List Learners in mlr3verse

Lists all learners, properties, and associated packages in a table that can be filtered and queried.

list_mlr3learners(select = NULL, filter = NULL)

Arguments

select
character()
Passed to data.table::subset.

filter
list()
Named list of conditions to filter on, names correspond to column names in table.

Examples

list_mlr3learners(
  select = c("id", "properties", "predict_types"),
  filter = list(class = "surv", predict_types = "distr"))

#>                   id                                    properties
#>  1:     surv.akritas                                              
#>  2:  surv.blackboost                                       weights
#>  3:     surv.cforest                                       weights
#>  4:    surv.coxboost                                       weights
#>  5:       surv.coxph                                       weights
#>  6:     surv.coxtime                                              
#>  7:       surv.ctree                                       weights
#>  8: surv.cv_coxboost                                       weights
#>  9:     surv.deephit                                              
#> 10:    surv.deepsurv                                              
#> 11:     surv.dnnsurv                                              
#> 12:    surv.flexible                                       weights
#> 13:    surv.gamboost          importance,selected_features,weights
#> 14:    surv.glmboost                                       weights
#> 15:      surv.kaplan                                      missings
#> 16:      surv.loghaz                                              
#> 17:      surv.mboost          importance,selected_features,weights
#> 18:      surv.nelson                                      missings
#> 19:  surv.obliqueRSF                            missings,oob_error
#> 20:  surv.parametric                                       weights
#> 21:    surv.pchazard                                              
#> 22:   surv.penalized                                              
#> 23:      surv.ranger                  importance,oob_error,weights
#> 24:       surv.rfsrc         importance,missings,oob_error,weights
#> 25:       surv.rpart importance,missings,selected_features,weights
#>                   id                                    properties
#>      predict_types
#>  1:    crank,distr
#>  2: distr,crank,lp
#>  3:    distr,crank
#>  4: distr,crank,lp
#>  5: distr,crank,lp
#>  6:    crank,distr
#>  7:    distr,crank
#>  8: distr,crank,lp
#>  9:    crank,distr
#> 10:    crank,distr
#> 11:    crank,distr
#> 12: distr,crank,lp
#> 13: distr,crank,lp
#> 14: distr,crank,lp
#> 15:    crank,distr
#> 16:    crank,distr
#> 17: distr,crank,lp
#> 18:    crank,distr
#> 19:    crank,distr
#> 20: distr,lp,crank
#> 21:    crank,distr
#> 22:    distr,crank
#> 23:    distr,crank
#> 24:    crank,distr
#> 25:    crank,distr
#>      predict_types