Classification C5.0 Learner

Calls C50::C5.0 from package C50.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.C50")
lrn("classif.C50")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “numeric”, “factor”, “ordered”

Required Packages: mlr3extralearners, C50

Parameters

Id	Type	Default	Levels	Range

trials	integer	1		\([1, \infty)\)

rules	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

subset	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

bands	integer	-		\([0, 1000]\)

winnow	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

noGlobalPruning	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

CF	numeric	0.25		\([0, 1]\)

minCases	integer	2		\([0, \infty)\)

fuzzyThreshold	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

sample	numeric	0		\([0, 0.999]\)

seed	integer	-		\((-\infty, \infty)\)

earlyStopping	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

label	list	outcome		\((-\infty, \infty)\)

na.action	list	function (object, ...) , object		\((-\infty, \infty)\)

References

Quinlan R (1993). C4.5: Programs for Machine Learning http://www.rulequest.com/see5-unix.html

Author

henrifnk

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifC50

Methods

Public methods

LearnerClassifC50$new()

LearnerClassifC50$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifC50$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifC50$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.