Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 R (0.45323741 0.54676259)  
#>   2) V12>=0.1941 75 22 M (0.70666667 0.29333333)  
#>     4) V34< 0.67925 65 12 M (0.81538462 0.18461538)  
#>       8) V43>=0.1883 49  2 M (0.95918367 0.04081633) *
#>       9) V43< 0.1883 16  6 R (0.37500000 0.62500000) *
#>     5) V34>=0.67925 10  0 R (0.00000000 1.00000000) *
#>   3) V12< 0.1941 64 10 R (0.15625000 0.84375000)  
#>     6) V46>=0.1639 9  2 M (0.77777778 0.22222222) *
#>     7) V46< 0.1639 55  3 R (0.05454545 0.94545455) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V49>=0.03625 84 17 M (0.79761905 0.20238095)  
#>      4) V37< 0.4634 61  2 M (0.96721311 0.03278689) *
#>      5) V37>=0.4634 23  8 R (0.34782609 0.65217391)  
#>       10) V49>=0.0755 9  1 M (0.88888889 0.11111111) *
#>       11) V49< 0.0755 14  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.03625 55  8 R (0.14545455 0.85454545) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 M (0.50359712 0.49640288)  
#>   2) V11>=0.21675 68 13 M (0.80882353 0.19117647)  
#>     4) V15< 0.645 56  5 M (0.91071429 0.08928571) *
#>     5) V15>=0.645 12  4 R (0.33333333 0.66666667) *
#>   3) V11< 0.21675 71 15 R (0.21126761 0.78873239)  
#>     6) V5>=0.099 13  3 M (0.76923077 0.23076923) *
#>     7) V5< 0.099 58  5 R (0.08620690 0.91379310) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V21>=0.6469 82 25 M (0.69512195 0.30487805)  
#>      4) V38>=0.2283 55  7 M (0.87272727 0.12727273)  
#>        8) V31< 0.5757 48  2 M (0.95833333 0.04166667) *
#>        9) V31>=0.5757 7  2 R (0.28571429 0.71428571) *
#>      5) V38< 0.2283 27  9 R (0.33333333 0.66666667)  
#>       10) V56< 0.00485 7  0 M (1.00000000 0.00000000) *
#>       11) V56>=0.00485 20  2 R (0.10000000 0.90000000) *
#>    3) V21< 0.6469 57 14 R (0.24561404 0.75438596)  
#>      6) V36< 0.21165 12  3 M (0.75000000 0.25000000) *
#>      7) V36>=0.21165 45  5 R (0.11111111 0.88888889)  
#>       14) V52>=0.0189 9  4 M (0.55555556 0.44444444) *
#>       15) V52< 0.0189 36  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V11>=0.15105 107 44 M (0.58878505 0.41121495)  
#>      4) V51>=0.0109 73 19 M (0.73972603 0.26027397)  
#>        8) V21>=0.5274 52  5 M (0.90384615 0.09615385)  
#>         16) V31>=0.21455 45  0 M (1.00000000 0.00000000) *
#>         17) V31< 0.21455 7  2 R (0.28571429 0.71428571) *
#>        9) V21< 0.5274 21  7 R (0.33333333 0.66666667)  
#>         18) V27>=0.83755 7  0 M (1.00000000 0.00000000) *
#>         19) V27< 0.83755 14  0 R (0.00000000 1.00000000) *
#>      5) V51< 0.0109 34  9 R (0.26470588 0.73529412)  
#>       10) V7< 0.087 8  0 M (1.00000000 0.00000000) *
#>       11) V7>=0.087 26  1 R (0.03846154 0.96153846) *
#>    3) V11< 0.15105 32  2 R (0.06250000 0.93750000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 M (0.61870504 0.38129496)  
#>    2) V32>=0.21445 108 28 M (0.74074074 0.25925926)  
#>      4) V36< 0.47405 73  6 M (0.91780822 0.08219178) *
#>      5) V36>=0.47405 35 13 R (0.37142857 0.62857143)  
#>       10) V45>=0.26365 11  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.26365 24  2 R (0.08333333 0.91666667) *
#>    3) V32< 0.21445 31  6 R (0.19354839 0.80645161)  
#>      6) V43>=0.2171 8  2 M (0.75000000 0.25000000) *
#>      7) V43< 0.2171 23  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V12>=0.1786 97 31 M (0.68041237 0.31958763)  
#>      4) V24>=0.57065 71 13 M (0.81690141 0.18309859)  
#>        8) V5>=0.04175 54  4 M (0.92592593 0.07407407) *
#>        9) V5< 0.04175 17  8 R (0.47058824 0.52941176) *
#>      5) V24< 0.57065 26  8 R (0.30769231 0.69230769)  
#>       10) V34>=0.3931 10  3 M (0.70000000 0.30000000) *
#>       11) V34< 0.3931 16  1 R (0.06250000 0.93750000) *
#>    3) V12< 0.1786 42  5 R (0.11904762 0.88095238)  
#>      6) V53< 0.00515 9  4 M (0.55555556 0.44444444) *
#>      7) V53>=0.00515 33  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V45>=0.19565 58 17 M (0.70689655 0.29310345)  
#>      4) V21>=0.54295 37  1 M (0.97297297 0.02702703) *
#>      5) V21< 0.54295 21  5 R (0.23809524 0.76190476)  
#>       10) V12>=0.1942 7  2 M (0.71428571 0.28571429) *
#>       11) V12< 0.1942 14  0 R (0.00000000 1.00000000) *
#>    3) V45< 0.19565 81 22 R (0.27160494 0.72839506)  
#>      6) V28>=0.9405 8  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.9405 73 14 R (0.19178082 0.80821918)  
#>       14) V52>=0.01455 25 11 R (0.44000000 0.56000000)  
#>         28) V11>=0.21325 9  0 M (1.00000000 0.00000000) *
#>         29) V11< 0.21325 16  2 R (0.12500000 0.87500000) *
#>       15) V52< 0.01455 48  3 R (0.06250000 0.93750000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.43165468 0.56834532)  
#>    2) V54>=0.0225 18  0 M (1.00000000 0.00000000) *
#>    3) V54< 0.0225 121 42 R (0.34710744 0.65289256)  
#>      6) V35< 0.36175 66 30 M (0.54545455 0.45454545)  
#>       12) V56< 0.00555 26  2 M (0.92307692 0.07692308) *
#>       13) V56>=0.00555 40 12 R (0.30000000 0.70000000)  
#>         26) V60>=0.00535 18  7 M (0.61111111 0.38888889) *
#>         27) V60< 0.00535 22  1 R (0.04545455 0.95454545) *
#>      7) V35>=0.36175 55  6 R (0.10909091 0.89090909)  
#>       14) V27>=0.9351 9  4 M (0.55555556 0.44444444) *
#>       15) V27< 0.9351 46  1 R (0.02173913 0.97826087) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V27< 0.50195 39  8 M (0.79487179 0.20512821)  
#>      4) V14>=0.14685 29  1 M (0.96551724 0.03448276) *
#>      5) V14< 0.14685 10  3 R (0.30000000 0.70000000) *
#>    3) V27>=0.50195 100 34 R (0.34000000 0.66000000)  
#>      6) V26>=0.9596 14  0 M (1.00000000 0.00000000) *
#>      7) V26< 0.9596 86 20 R (0.23255814 0.76744186)  
#>       14) V43>=0.3442 15  5 M (0.66666667 0.33333333) *
#>       15) V43< 0.3442 71 10 R (0.14084507 0.85915493)  
#>         30) V50< 0.0086 8  2 M (0.75000000 0.25000000) *
#>         31) V50>=0.0086 63  4 R (0.06349206 0.93650794)  
#>           62) V54>=0.0169 7  3 M (0.57142857 0.42857143) *
#>           63) V54< 0.0169 56  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.6450753 0.6224661 0.5576063 0.7047701 0.7808891 0.7539233 0.6294120
#>  [8] 0.7603183 0.6346207 0.6934969
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.5142416 5.2683365
#>   [2,] 2.0698863 4.7126918
#>   [3,] 2.1593168 4.6232612
#>   [4,] 2.1593168 4.6232612
#>   [5,] 3.3999656 3.3826125
#>   [6,] 1.9731162 4.8094619
#>   [7,] 2.6521068 4.1304713
#>   [8,] 2.7157803 4.0667978
#>   [9,] 2.1521904 4.6303877
#>  [10,] 2.1521904 4.6303877
#>  [11,] 2.0328878 4.7496903
#>  [12,] 1.8961785 4.8863996
#>  [13,] 2.6521068 4.1304713
#>  [14,] 2.5904941 4.1920840
#>  [15,] 1.9575296 4.8250485
#>  [16,] 2.5194633 4.2631148
#>  [17,] 1.9473366 4.8352415
#>  [18,] 1.8805152 4.9020629
#>  [19,] 2.6508767 4.1317014
#>  [20,] 1.2744873 5.5080908
#>  [21,] 1.3989986 5.3835795
#>  [22,] 1.3833353 5.3992427
#>  [23,] 2.0432656 4.7393125
#>  [24,] 1.8857240 4.8968541
#>  [25,] 0.6450753 6.1375028
#>  [26,] 2.2395825 4.5429955
#>  [27,] 2.0328878 4.7496903
#>  [28,] 1.9610382 4.8215399
#>  [29,] 1.4474202 5.3351579
#>  [30,] 2.6573155 4.1252626
#>  [31,] 0.0000000 6.7825781
#>  [32,] 1.3229089 5.4596692
#>  [33,] 1.4474202 5.3351579
#>  [34,] 1.3159629 5.4666152
#>  [35,] 0.7539233 6.0286548
#>  [36,] 1.8805152 4.9020629
#>  [37,] 1.9473366 4.8352415
#>  [38,] 2.0433423 4.7392358
#>  [39,] 3.2303606 3.5522175
#>  [40,] 2.0110102 4.7715679
#>  [41,] 0.6294120 6.1531661
#>  [42,] 1.3281176 5.4544605
#>  [43,] 0.0000000 6.7825781
#>  [44,] 1.3949390 5.3876391
#>  [45,] 0.6294120 6.1531661
#>  [46,] 0.6294120 6.1531661
#>  [47,] 0.6294120 6.1531661
#>  [48,] 1.3229089 5.4596692
#>  [49,] 0.6294120 6.1531661
#>  [50,] 1.9575296 4.8250485
#>  [51,] 2.0328878 4.7496903
#>  [52,] 2.0820410 4.7005371
#>  [53,] 1.4474202 5.3351579
#>  [54,] 0.7539233 6.0286548
#>  [55,] 1.3763894 5.4061887
#>  [56,] 0.7539233 6.0286548
#>  [57,] 1.9731162 4.8094619
#>  [58,] 2.7866905 3.9958876
#>  [59,] 1.9339957 4.8485823
#>  [60,] 2.6508767 4.1317014
#>  [61,] 2.0484304 4.7341477
#>  [62,] 1.2511032 5.5314749
#>  [63,] 2.0276790 4.7548991
#>  [64,] 0.7603183 6.0222598
#>  [65,] 2.0162998 4.7662783
#>  [66,] 0.7539233 6.0286548
#>  [67,] 2.0933142 4.6892639
#>  [68,] 3.9083748 2.8742033
#>  [69,] 5.3566138 1.4259643
#>  [70,] 5.5028821 1.2796960
#>  [71,] 4.7547185 2.0278596
#>  [72,] 4.8804160 1.9021621
#>  [73,] 6.1601120 0.6224661
#>  [74,] 5.3670683 1.4155098
#>  [75,] 5.3997938 1.3827843
#>  [76,] 4.6119588 2.1706193
#>  [77,] 5.2413708 1.5412073
#>  [78,] 6.0222598 0.7603183
#>  [79,] 6.7825781 0.0000000
#>  [80,] 4.7446022 2.0379759
#>  [81,] 5.3835795 1.3989986
#>  [82,] 6.0778079 0.7047701
#>  [83,] 5.4544605 1.3281176
#>  [84,] 5.5903510 1.1922271
#>  [85,] 5.3174897 1.4650884
#>  [86,] 6.0890812 0.6934969
#>  [87,] 5.5150368 1.2675413
#>  [88,] 4.8863996 1.8961785
#>  [89,] 4.1304713 2.6521068
#>  [90,] 5.4646535 1.3179246
#>  [91,] 4.8146707 1.9679074
#>  [92,] 6.1375028 0.6450753
#>  [93,] 4.9574304 1.8251477
#>  [94,] 5.3722770 1.4103011
#>  [95,] 5.5254913 1.2570868
#>  [96,] 5.5185453 1.2640328
#>  [97,] 5.2413708 1.5412073
#>  [98,] 5.5955597 1.1870184
#>  [99,] 6.2249717 0.5576063
#> [100,] 4.8137752 1.9688029
#> [101,] 6.1531661 0.6294120
#> [102,] 4.7393125 2.0432656
#> [103,] 3.5036153 3.2789628
#> [104,] 4.6836877 2.0988904
#> [105,] 5.5185453 1.2640328
#> [106,] 4.8216166 1.9609615
#> [107,] 4.7990074 1.9835707
#> [108,] 6.0778079 0.7047701
#> [109,] 5.4431872 1.3393909
#> [110,] 6.1479573 0.6346207
#> [111,] 6.0778079 0.7047701
#> [112,] 6.1479573 0.6346207
#> [113,] 4.0377937 2.7447844
#> [114,] 4.0491437 2.7334344
#> [115,] 4.8216166 1.9609615
#> [116,] 5.6025057 1.1800724
#> [117,] 6.1601120 0.6224661
#> [118,] 4.6950236 2.0875545
#> [119,] 5.3792230 1.4033551
#> [120,] 4.7703818 2.0121963
#> [121,] 4.9504845 1.8320936
#> [122,] 6.0778079 0.7047701
#> [123,] 6.0890812 0.6934969
#> [124,] 6.1479573 0.6346207
#> [125,] 6.0778079 0.7047701
#> [126,] 6.7825781 0.0000000
#> [127,] 5.3722770 1.4103011
#> [128,] 5.3670683 1.4155098
#> [129,] 6.1531661 0.6294120
#> [130,] 4.8093852 1.9731929
#> [131,] 5.4440060 1.3385721
#> [132,] 5.5028821 1.2796960
#> [133,] 5.3722770 1.4103011
#> [134,] 6.7825781 0.0000000
#> [135,] 5.4431872 1.3393909
#> [136,] 6.7825781 0.0000000
#> [137,] 6.0890812 0.6934969
#> [138,] 5.3566138 1.4259643
#> [139,] 5.5307000 1.2518781
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.22325458 0.77674542
#>   [2,] 0.30517692 0.69482308
#>   [3,] 0.31836225 0.68163775
#>   [4,] 0.31836225 0.68163775
#>   [5,] 0.50127924 0.49872076
#>   [6,] 0.29090946 0.70909054
#>   [7,] 0.39101750 0.60898250
#>   [8,] 0.40040531 0.59959469
#>   [9,] 0.31731155 0.68268845
#>  [10,] 0.31731155 0.68268845
#>  [11,] 0.29972198 0.70027802
#>  [12,] 0.27956604 0.72043396
#>  [13,] 0.39101750 0.60898250
#>  [14,] 0.38193355 0.61806645
#>  [15,] 0.28861144 0.71138856
#>  [16,] 0.37146101 0.62853899
#>  [17,] 0.28710862 0.71289138
#>  [18,] 0.27725670 0.72274330
#>  [19,] 0.39083614 0.60916386
#>  [20,] 0.18790602 0.81209398
#>  [21,] 0.20626354 0.79373646
#>  [22,] 0.20395421 0.79604579
#>  [23,] 0.30125205 0.69874795
#>  [24,] 0.27802466 0.72197534
#>  [25,] 0.09510768 0.90489232
#>  [26,] 0.33019635 0.66980365
#>  [27,] 0.29972198 0.70027802
#>  [28,] 0.28912873 0.71087127
#>  [29,] 0.21340266 0.78659734
#>  [30,] 0.39178546 0.60821454
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.19504514 0.80495486
#>  [33,] 0.21340266 0.78659734
#>  [34,] 0.19402105 0.80597895
#>  [35,] 0.11115587 0.88884413
#>  [36,] 0.27725670 0.72274330
#>  [37,] 0.28710862 0.71289138
#>  [38,] 0.30126336 0.69873664
#>  [39,] 0.47627327 0.52372673
#>  [40,] 0.29649642 0.70350358
#>  [41,] 0.09279834 0.90720166
#>  [42,] 0.19581310 0.80418690
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.20566501 0.79433499
#>  [45,] 0.09279834 0.90720166
#>  [46,] 0.09279834 0.90720166
#>  [47,] 0.09279834 0.90720166
#>  [48,] 0.19504514 0.80495486
#>  [49,] 0.09279834 0.90720166
#>  [50,] 0.28861144 0.71138856
#>  [51,] 0.29972198 0.70027802
#>  [52,] 0.30696896 0.69303104
#>  [53,] 0.21340266 0.78659734
#>  [54,] 0.11115587 0.88884413
#>  [55,] 0.20293012 0.79706988
#>  [56,] 0.11115587 0.88884413
#>  [57,] 0.29090946 0.70909054
#>  [58,] 0.41086007 0.58913993
#>  [59,] 0.28514169 0.71485831
#>  [60,] 0.39083614 0.60916386
#>  [61,] 0.30201353 0.69798647
#>  [62,] 0.18445836 0.81554164
#>  [63,] 0.29895403 0.70104597
#>  [64,] 0.11209871 0.88790129
#>  [65,] 0.29727632 0.70272368
#>  [66,] 0.11115587 0.88884413
#>  [67,] 0.30863105 0.69136895
#>  [68,] 0.57623735 0.42376265
#>  [69,] 0.78976072 0.21023928
#>  [70,] 0.81132602 0.18867398
#>  [71,] 0.70101936 0.29898064
#>  [72,] 0.71955176 0.28044824
#>  [73,] 0.90822574 0.09177426
#>  [74,] 0.79130210 0.20869790
#>  [75,] 0.79612703 0.20387297
#>  [76,] 0.67997135 0.32002865
#>  [77,] 0.77276969 0.22723031
#>  [78,] 0.88790129 0.11209871
#>  [79,] 1.00000000 0.00000000
#>  [80,] 0.69952784 0.30047216
#>  [81,] 0.79373646 0.20626354
#>  [82,] 0.89609111 0.10390889
#>  [83,] 0.80418690 0.19581310
#>  [84,] 0.82422214 0.17577786
#>  [85,] 0.78399240 0.21600760
#>  [86,] 0.89775320 0.10224680
#>  [87,] 0.81311807 0.18688193
#>  [88,] 0.72043396 0.27956604
#>  [89,] 0.60898250 0.39101750
#>  [90,] 0.80568973 0.19431027
#>  [91,] 0.70985849 0.29014151
#>  [92,] 0.90489232 0.09510768
#>  [93,] 0.73090650 0.26909350
#>  [94,] 0.79207006 0.20792994
#>  [95,] 0.81465944 0.18534056
#>  [96,] 0.81363536 0.18636464
#>  [97,] 0.77276969 0.22723031
#>  [98,] 0.82499009 0.17500991
#>  [99,] 0.91778844 0.08221156
#> [100,] 0.70972647 0.29027353
#> [101,] 0.90720166 0.09279834
#> [102,] 0.69874795 0.30125205
#> [103,] 0.51656100 0.48343900
#> [104,] 0.69054682 0.30945318
#> [105,] 0.81363536 0.18636464
#> [106,] 0.71088258 0.28911742
#> [107,] 0.70754916 0.29245084
#> [108,] 0.89609111 0.10390889
#> [109,] 0.80252481 0.19747519
#> [110,] 0.90643370 0.09356630
#> [111,] 0.89609111 0.10390889
#> [112,] 0.90643370 0.09356630
#> [113,] 0.59531842 0.40468158
#> [114,] 0.59699183 0.40300817
#> [115,] 0.71088258 0.28911742
#> [116,] 0.82601418 0.17398582
#> [117,] 0.90822574 0.09177426
#> [118,] 0.69221815 0.30778185
#> [119,] 0.79309414 0.20690586
#> [120,] 0.70332869 0.29667131
#> [121,] 0.72988242 0.27011758
#> [122,] 0.89609111 0.10390889
#> [123,] 0.89775320 0.10224680
#> [124,] 0.90643370 0.09356630
#> [125,] 0.89609111 0.10390889
#> [126,] 1.00000000 0.00000000
#> [127,] 0.79207006 0.20792994
#> [128,] 0.79130210 0.20869790
#> [129,] 0.90720166 0.09279834
#> [130,] 0.70907923 0.29092077
#> [131,] 0.80264553 0.19735447
#> [132,] 0.81132602 0.18867398
#> [133,] 0.79207006 0.20792994
#> [134,] 1.00000000 0.00000000
#> [135,] 0.80252481 0.19747519
#> [136,] 1.00000000 0.00000000
#> [137,] 0.89775320 0.10224680
#> [138,] 0.78976072 0.21023928
#> [139,] 0.81542740 0.18457260
#> 
#> $class
#>   [1] "R" "R" "R" "R" "M" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 0.0000000 9.2245192 8.5004977 0.0000000 1.3508641 1.0864960 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 8.3342082 0.0000000 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 1.8369716 0.0000000 2.9085539 5.3415736 2.1178573 0.0000000 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 2.5794992 3.2158296 0.0000000 3.1293955 2.1441135 4.7610657 2.3590298 2.1962783 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 4.4635978 0.0000000 5.7060351 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 1.5434992 0.0000000 0.0000000 6.7979801 2.6318418 1.3722176 2.4173890 1.9797199 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.8124793 3.3403624 0.0000000 4.0468609 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 1.1889683 2.6122951 0.0000000 0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V11       V12       V21       V49       V45       V27       V36       V43 
#> 9.2245192 8.5004977 8.3342082 6.7979801 5.7060351 5.3415736 4.7610657 4.4635978 
#>       V56       V54       V32       V34       V26        V5        V7       V31 
#> 4.0468609 3.3403624 3.2158296 3.1293955 2.9085539 2.6318418 2.6122951 2.5794992 
#>       V51       V37       V38       V35       V28       V52       V24       V46 
#> 2.4173890 2.3590298 2.1962783 2.1441135 2.1178573 1.9797199 1.8369716 1.5434992 
#>       V50       V14       V60       V15       V53        V1       V10       V13 
#> 1.3722176 1.3508641 1.1889683 1.0864960 0.8124793 0.0000000 0.0000000 0.0000000 
#>       V16       V17       V18       V19        V2       V20       V22       V23 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V25       V29        V3       V30       V33       V39        V4       V40 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V41       V42       V44       V47       V48       V55       V57       V58 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V59        V6        V8        V9 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913