Oblique Random Forest Classifier

Accelerated oblique random classification forest. Calls aorsf::orsf() from aorsf. Note that although the learner has the property "missing" and it can in principle deal with missing values, the behaviour has to be configured using the parameter na_action.

Initial parameter values

n_thread: This parameter is initialized to 1 (default is 0) to avoid conflicts with the mlr3 parallelization.

pred_simplify has to be TRUE, otherwise response is NA in prediction

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.aorsf")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, aorsf

Parameters

Id	Type	Default	Levels	Range

attach_data	logical	TRUE	TRUE, FALSE	-

epsilon	numeric	1e-09		\([0, \infty)\)

importance	character	anova	none, anova, negate, permute	-

importance_max_pvalue	numeric	0.01		\([1e-04, 0.9999]\)

leaf_min_events	integer	1		\([1, \infty)\)

leaf_min_obs	integer	5		\([1, \infty)\)

max_iter	integer	20		\([1, \infty)\)

method	character	glm	glm, net, pca, random	-

mtry	integer	NULL		\([1, \infty)\)

mtry_ratio	numeric	-		\([0, 1]\)

n_retry	integer	3		\([0, \infty)\)

n_split	integer	5		\([1, \infty)\)

n_thread	integer	-		\([0, \infty)\)

n_tree	integer	500		\([1, \infty)\)

na_action	character	fail	fail, impute_meanmode	-

net_mix	numeric	0.5		\((-\infty, \infty)\)

oobag	logical	FALSE	TRUE, FALSE	-

oobag_eval_every	integer	NULL		\([1, \infty)\)

oobag_fun	untyped	NULL		-

oobag_pred_type	character	prob	none, leaf, prob, class	-

pred_aggregate	logical	TRUE	TRUE, FALSE	-

sample_fraction	numeric	0.632		\([0, 1]\)

sample_with_replacement	logical	TRUE	TRUE, FALSE	-

scale_x	logical	FALSE	TRUE, FALSE	-

split_min_events	integer	5		\([1, \infty)\)

split_min_obs	integer	10		\([1, \infty)\)

split_min_stat	numeric	NULL		\([0, \infty)\)

split_rule	character	gini	gini, cstat	-

target_df	integer	NULL		\([1, \infty)\)

tree_seeds	integer	NULL		\([1, \infty)\)

verbose_progress	logical	FALSE	TRUE, FALSE	-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifObliqueRandomForest

Methods

Public methods

LearnerClassifObliqueRandomForest$new()

LearnerClassifObliqueRandomForest$oob_error()

LearnerClassifObliqueRandomForest$importance()

LearnerClassifObliqueRandomForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifObliqueRandomForest$new()

Method oob_error()

OOB concordance error extracted from the model slot eval_oobag$stat_values

Usage

LearnerClassifObliqueRandomForest$oob_error()

Returns

numeric().

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifObliqueRandomForest$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifObliqueRandomForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.aorsf")
print(learner)
#> 
#> ── <LearnerClassifObliqueRandomForest> (classif.aorsf): Oblique Random Forest Cl
#> • Model: -
#> • Parameters: n_thread=1
#> • Packages: mlr3, mlr3extralearners, and aorsf
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, oob_error, twoclass, and
#> weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> ---------- Oblique random classification forest
#> 
#>      Linear combinations: Logistic regression
#>           N observations: 139
#>                N classes: 2
#>                  N trees: 500
#>       N predictors total: 60
#>    N predictors per node: 8
#>  Average leaves per tree: 4.908
#> Min observations in leaf: 5
#>           OOB stat value: 0.88
#>            OOB stat type: AUC-ROC
#>      Variable importance: anova
#> 
#> -----------------------------------------
print(learner$importance())
#>         V11         V52         V51         V12         V36          V9 
#> 0.451754386 0.391509434 0.364055300 0.353773585 0.351111111 0.347368421 
#>         V10         V37         V35         V49         V31         V48 
#> 0.324324324 0.280193237 0.250000000 0.229729730 0.205882353 0.195348837 
#>          V1         V45         V13         V44         V47         V20 
#> 0.188172043 0.180000000 0.172248804 0.167464115 0.144736842 0.137254902 
#>         V46          V5         V21         V34         V19         V16 
#> 0.136792453 0.133663366 0.132812500 0.122881356 0.116751269 0.103896104 
#>         V43          V4         V41         V28          V8         V22 
#> 0.097777778 0.096491228 0.088495575 0.075117371 0.071748879 0.067567568 
#>          V7         V14         V50         V32         V57         V42 
#> 0.063025210 0.062500000 0.061611374 0.061032864 0.059907834 0.059907834 
#>         V33         V54         V40         V15         V17         V58 
#> 0.058558559 0.058536585 0.056521739 0.050458716 0.049792531 0.049180328 
#>         V25         V29         V23         V27         V38         V39 
#> 0.045662100 0.045248869 0.043478261 0.042654028 0.042452830 0.041152263 
#>          V3         V59         V55         V26          V2         V30 
#> 0.040723982 0.038834951 0.038461538 0.038251366 0.037037037 0.036866359 
#>         V53         V18         V24          V6         V56         V60 
#> 0.033816425 0.032710280 0.030837004 0.018264840 0.016528926 0.008849558 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058