BlockForest Classification Learner

Random forests for blocks of clinical and omics covariate data. Calls blockForest::blockfor() from package blockForest.

In this learner, only the trained forest object ($forest) is retained. The optimized block-specific tuning parameters (paramvalues) and the biased OOB error estimate (biased_oob_error_donotuse) are discarded, as they are either not needed for downstream use or not reliable for performance estimation.

Initial parameter values

num.threads is initialized to 1 to avoid conflicts with parallelization via future.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.blockforest")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, blockForest

Parameters

Id	Type	Default	Levels	Range

blocks	untyped	-		-

block.method	character	BlockForest	BlockForest, RandomBlock, BlockVarSel, VarProb, SplitWeights	-

num.trees	integer	2000		\([1, \infty)\)

mtry	untyped	NULL		-

nsets	integer	300		\([1, \infty)\)

num.trees.pre	integer	1500		\([1, \infty)\)

splitrule	character	extratrees	extratrees, gini	-

always.select.block	integer	0		\([0, 1]\)

importance	character	-	none, impurity, impurity_corrected, permutation	-

num.threads	integer	-		\([1, \infty)\)

seed	integer	NULL		\((-\infty, \infty)\)

verbose	logical	TRUE	TRUE, FALSE	-

References

Hornung, R., Wright, N. M (2019). “Block Forests: Random forests for blocks of clinical and omics covariate data.” BMC Bioinformatics, 20(1), 1–17. doi:10.1186/s12859-019-2942-y , https://doi.org/10.1186/s12859-019-2942-y.

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifBlockForest

Methods

Public methods

LearnerClassifBlockForest$new()

LearnerClassifBlockForest$importance()

LearnerClassifBlockForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifBlockForest$new()

Method importance()

The importance scores are extracted from the model slot variable.importance.

Usage

LearnerClassifBlockForest$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifBlockForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define a Task
task = tsk("sonar")
# Create train and test set
ids = partition(task)
# check task's features
task$feature_names
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
# partition features to 2 blocks
blocks = list(bl1 = 1:42, bl2 = 43:60)
# define learner
learner = lrn("classif.blockforest", blocks = blocks,
              importance = "permutation", nsets = 10, predict_type = "prob",
              num.trees = 50, num.trees.pre = 10, splitrule = "gini")
# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
# feature importance
learner$importance()
#>           V11           V12           V36           V49            V9 
#>  3.167134e-02  1.652017e-02  1.096023e-02  1.049497e-02  1.035416e-02 
#>           V47           V45           V48           V35           V46 
#>  7.628520e-03  6.942053e-03  6.858749e-03  6.036699e-03  5.547907e-03 
#>           V37           V52           V21           V34           V22 
#>  4.546695e-03  4.347189e-03  3.961405e-03  3.792172e-03  3.772366e-03 
#>           V28           V51           V18           V39            V5 
#>  3.651450e-03  3.388433e-03  3.188724e-03  3.162243e-03  3.070339e-03 
#>           V58           V23           V13            V4            V8 
#>  2.826567e-03  2.739181e-03  2.646820e-03  2.558916e-03  2.531492e-03 
#>            V2           V15           V17            V3           V53 
#>  2.423667e-03  2.233711e-03  2.208946e-03  2.070916e-03  2.040131e-03 
#>           V44           V38           V50           V16           V31 
#>  1.971372e-03  1.908079e-03  1.889674e-03  1.817823e-03  1.757238e-03 
#>           V43           V56           V57           V10           V19 
#>  1.586237e-03  1.541309e-03  1.532537e-03  1.493382e-03  1.429831e-03 
#>           V24           V55           V32           V30           V25 
#>  1.349366e-03  1.309730e-03  1.221359e-03  1.063240e-03  8.947378e-04 
#>           V29            V6           V54           V26           V60 
#>  8.902842e-04  7.944147e-04  5.214672e-04  4.873966e-04  1.232228e-04 
#>           V42           V41           V33           V20            V7 
#>  9.616214e-05  8.686525e-05  8.474520e-05 -9.615385e-05 -3.162408e-04 
#>           V27            V1           V14           V59           V40 
#> -5.014745e-04 -7.308686e-04 -1.020548e-03 -2.021112e-03 -3.035660e-03 
# Make predictions for the test observations
pred = learner$predict(task, row_ids = ids$test)
pred
#> 
#> ── <PredictionClassif> for 69 observations: ────────────────────────────────────
#>  row_ids truth response    prob.M    prob.R
#>        2     R        M 0.5054603 0.4945397
#>        3     R        M 0.5811667 0.4188333
#>        4     R        R 0.4502063 0.5497937
#>      ---   ---      ---       ---       ---
#>      202     M        M 0.8347222 0.1652778
#>      207     M        M 0.6694921 0.3305079
#>      208     M        R 0.4846667 0.5153333
# Score the predictions
pred$score()
#> classif.ce 
#>  0.1594203