Gradient Boosted Decision Trees Classification Learner

Calls catboost::catboost.train from package catboost.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.catboost")
lrn("classif.catboost")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “numeric”, “factor”, “ordered”

Required Packages: mlr3extralearners, catboost

Parameters

Id	Type	Default	Levels	Range

loss_function_twoclass	character	Logloss	Logloss, CrossEntropy	\((-\infty, \infty)\)

loss_function_multiclass	character	MultiClass	MultiClass, MultiClassOneVsAll	\((-\infty, \infty)\)

iterations	integer	1000		\([1, \infty)\)

learning_rate	numeric	0.03		\([0.001, 1]\)

random_seed	integer	0		\([0, \infty)\)

l2_leaf_reg	numeric	3		\([0, \infty)\)

bootstrap_type	character	-	Bayesian, Bernoulli, MVS, Poisson, No	\((-\infty, \infty)\)

bagging_temperature	numeric	1		\([0, \infty)\)

subsample	numeric	-		\([0, 1]\)

sampling_frequency	character	PerTreeLevel	PerTree, PerTreeLevel	\((-\infty, \infty)\)

sampling_unit	character	Object	Object, Group	\((-\infty, \infty)\)

mvs_reg	numeric	-		\([0, \infty)\)

random_strength	numeric	1		\([0, \infty)\)

depth	integer	6		\([1, 16]\)

grow_policy	character	SymmetricTree	SymmetricTree, Depthwise, Lossguide	\((-\infty, \infty)\)

min_data_in_leaf	integer	1		\([1, \infty)\)

max_leaves	integer	31		\([1, \infty)\)

ignored_features	list	NULL		\((-\infty, \infty)\)

one_hot_max_size	list	FALSE		\((-\infty, \infty)\)

has_time	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

rsm	numeric	1		\([0.001, 1]\)

nan_mode	character	Min	Min, Max	\((-\infty, \infty)\)

fold_permutation_block	integer	-		\([1, 256]\)

leaf_estimation_method	character	-	Newton, Gradient, Exact	\((-\infty, \infty)\)

leaf_estimation_iterations	integer	-		\([1, \infty)\)

leaf_estimation_backtracking	character	AnyImprovement	No, AnyImprovement, Armijo	\((-\infty, \infty)\)

fold_len_multiplier	numeric	2		\([1.001, \infty)\)

approx_on_full_history	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

class_weights	list	-		\((-\infty, \infty)\)

auto_class_weights	character	None	None, Balanced, SqrtBalanced	\((-\infty, \infty)\)

boosting_type	character	-	Ordered, Plain	\((-\infty, \infty)\)

boost_from_average	logical	-	TRUE, FALSE	\((-\infty, \infty)\)

langevin	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

diffusion_temperature	numeric	10000		\([0, \infty)\)

score_function	character	Cosine	Cosine, L2, NewtonCosine, NewtonL2	\((-\infty, \infty)\)

monotone_constraints	list	-		\((-\infty, \infty)\)

feature_weights	list	-		\((-\infty, \infty)\)

first_feature_use_penalties	list	-		\((-\infty, \infty)\)

penalties_coefficient	numeric	1		\([0, \infty)\)

per_object_feature_penalties	list	-		\((-\infty, \infty)\)

model_shrink_rate	numeric	-		\((-\infty, \infty)\)

model_shrink_mode	character	-	Constant, Decreasing	\((-\infty, \infty)\)

target_border	numeric	-		\((-\infty, \infty)\)

border_count	integer	-		\([1, 65535]\)

feature_border_type	character	GreedyLogSum	Median, Uniform, UniformAndQuantiles, MaxLogSum, MinEntropy, GreedyLogSum	\((-\infty, \infty)\)

per_float_feature_quantization	list	-		\((-\infty, \infty)\)

classes_count	integer	-		\([1, \infty)\)

thread_count	integer	1		\([-1, \infty)\)

task_type	character	CPU	CPU, GPU	\((-\infty, \infty)\)

devices	list	-		\((-\infty, \infty)\)

logging_level	character	Silent	Silent, Verbose, Info, Debug	\((-\infty, \infty)\)

metric_period	integer	1		\([1, \infty)\)

train_dir	list	catboost_info		\((-\infty, \infty)\)

model_size_reg	numeric	0.5		\([0, 1]\)

allow_writing_files	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

save_snapshot	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

snapshot_file	list	-		\((-\infty, \infty)\)

snapshot_interval	integer	600		\([1, \infty)\)

simple_ctr	list	-		\((-\infty, \infty)\)

combinations_ctr	list	-		\((-\infty, \infty)\)

ctr_target_border_count	integer	-		\([1, 255]\)

counter_calc_method	character	Full	SkipTest, Full	\((-\infty, \infty)\)

max_ctr_complexity	integer	-		\([1, \infty)\)

ctr_leaf_count_limit	integer	-		\([1, \infty)\)

store_all_simple_ctr	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

final_ctr_computation_mode	character	Default	Default, Skip	\((-\infty, \infty)\)

verbose	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

ntree_start	integer	0		\([0, \infty)\)

ntree_end	integer	0		\([0, \infty)\)

Installation

The easiest way to install catboost is with the helper function install_catboost.

Custom mlr3 defaults

logging_level:

Actual default: "Verbose"

Adjusted default: "Silent"

Reason for change: consistent with other mlr3 learners

thread_count:

Actual default: -1

Adjusted default: 1

Reason for change: consistent with other mlr3 learners

allow_writing_files:

Actual default: TRUE

Adjusted default: FALSE

Reason for change: consistent with other mlr3 learners

save_snapshot:

Actual default: TRUE

Adjusted default: FALSE

Reason for change: consistent with other mlr3 learners

References

catboost: unbiased boosting with categorical features. Liudmila Prokhorenkova, Gleb Guse, Aleksandr Vorobev, Anna Veronika Dorogush and Andrey Gulin. 2017. https://arxiv.org/abs/1706.09516.

catboost: gradient boosting with categorical features support. Anna Veronika Dorogush, Vasily Ershov and Andrey Gulin. 2018. https://arxiv.org/abs/1810.11363.

See also

Author

sumny

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifCatboost

Methods

Public methods

LearnerClassifCatboost$new()

LearnerClassifCatboost$importance()

LearnerClassifCatboost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Create a LearnerClassifCatboost object.

Usage

LearnerClassifCatboost$new()

Method importance()

The importance scores are calculated using catboost.get_feature_importance, setting type = "FeatureImportance", returned for 'all'.

Usage

LearnerClassifCatboost$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifCatboost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("catboost", quietly = TRUE)) {
  learner = mlr3::lrn("classif.catboost")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}