Classification Conditional Inference Tree Learner

Calls partykit::ctree from package partykit.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.ctree")
lrn("classif.ctree")

Traits

Packages: partykit, sandwich, coin

Predict Types: response, prob

Feature Types: integer, numeric, factor, ordered

Properties: multiclass, twoclass, weights

References

Hothorn T, Zeileis A (2015). “partykit: A Modular Toolkit for Recursive Partytioning in R.” Journal of Machine Learning Research, 16(118), 3905-3909. http://jmlr.org/papers/v16/hothorn15a.html

Hothorn T, Hornik K, Zeileis A (2006). “Unbiased Recursive Partitioning: A Conditional Inference Framework.” Journal of Computational and Graphical Statistics, 15(3), 651–674. doi: 10.1198/106186006x133933

See also

Author

sumny

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifCTree

Methods

Public methods

LearnerClassifCTree$new()

LearnerClassifCTree$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifCTree$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifCTree$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("classif.ctree"))
print(learner)

#> <LearnerClassifCTree:classif.ctree>
#> * Model: -
#> * Parameters: list()
#> * Packages: partykit, sandwich, coin
#> * Predict Type: response
#> * Feature types: integer, numeric, factor, ordered
#> * Properties: multiclass, twoclass, weights

# available parameters:
learner$param_set$ids()

#>  [1] "teststat"        "splitstat"       "splittest"       "testtype"       
#>  [5] "nmax"            "alpha"           "mincriterion"    "logmincriterion"
#>  [9] "minsplit"        "minbucket"       "minprob"         "stump"          
#> [13] "lookahead"       "MIA"             "nresample"       "tol"            
#> [17] "maxsurrogate"    "numsurrogate"    "mtry"            "maxdepth"       
#> [21] "multiway"        "splittry"        "intersplit"      "majority"       
#> [25] "caseweights"     "maxvar"          "applyfun"        "cores"          
#> [29] "saveinfo"        "update"          "splitflavour"    "offset"         
#> [33] "cluster"         "scores"          "doFit"           "maxpts"         
#> [37] "abseps"          "releps"