Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.3948207 -0.15932808 -0.15275071 -0.142599452 -0.08750632 -0.202620261
#>  [2,] -0.2223688 -0.11206022 -0.10694604 -0.090650921 -0.02827552 -0.117394676
#>  [3,] -0.3268550 -0.06272350 -0.03433789 -0.044590693 -0.06366754  0.119836672
#>  [4,] -0.2095641 -0.12829072 -0.01977932 -0.094781509 -0.04192708 -0.042786014
#>  [5,] -0.2513450 -0.16740074 -0.10521427 -0.028490027 -0.02393113  0.009845302
#>  [6,] -0.2072703 -0.19600439 -0.12629844 -0.202659369 -0.21102457 -0.073199667
#>  [7,] -0.3518417  0.01745920 -0.02336866 -0.008905108 -0.05399222 -0.010914346
#>  [8,] -0.2529261 -0.03969432 -0.05453959 -0.035423979 -0.09204421  0.022876630
#>  [9,] -0.4219337 -0.22206499 -0.14676880 -0.042021834 -0.16871325 -0.159356161
#> [10,] -0.3122270 -0.09399582 -0.09045555 -0.167619757 -0.05060501 -0.070651545
#>                V15         V16          V17          V18          V19
#>  [1,] -0.099742524 -0.02540156  0.003458173 -0.004547363 -0.098974663
#>  [2,] -0.093986836 -0.14370166 -0.038529580  0.036906621 -0.089540665
#>  [3,]  0.098735269  0.10531566  0.134793392  0.036234530  0.054255762
#>  [4,] -0.076931419  0.06349705 -0.103582637 -0.043064575  0.001231629
#>  [5,] -0.026254117 -0.12408686 -0.098029740 -0.116917630  0.011184406
#>  [6,]  0.007701446 -0.14451589  0.042524588  0.011639585  0.068150569
#>  [7,] -0.037806143 -0.09753356  0.043002262  0.087832314  0.105390064
#>  [8,] -0.006021193  0.18097551  0.107048364  0.159325163  0.145332238
#>  [9,] -0.145419108 -0.01324448 -0.014737779 -0.019928765 -0.033951790
#> [10,] -0.067789181  0.07090546 -0.018156200 -0.040904637  0.170281766
#>               V2          V20          V21         V22          V23         V24
#>  [1,] -0.2402119  0.024212757  0.169126359 -0.00593046  0.103929209  0.11539759
#>  [2,] -0.3463517 -0.152329947 -0.166792999 -0.03256286 -0.114155166 -0.13435624
#>  [3,] -0.3811558  0.162838086  0.196959505  0.08651587  0.169654523  0.22374515
#>  [4,] -0.2519565  0.007055932  0.025413253  0.16275095  0.156887488  0.10377212
#>  [5,] -0.3233925  0.095994243  0.103986003  0.08710768  0.002237936  0.04750151
#>  [6,] -0.2699236  0.059689406  0.113212493  0.03408053  0.079540654  0.11717874
#>  [7,] -0.3199809  0.088096724  0.002557722  0.11658766  0.033976300 -0.04592630
#>  [8,] -0.2074093  0.169759496  0.094991402  0.19247586  0.031477981  0.05661739
#>  [9,] -0.3719159  0.114343799  0.132357603  0.07849560  0.164284384  0.20161195
#> [10,] -0.1768804  0.051160859  0.006835595 -0.01368679  0.172007063  0.09248752
#>               V25         V26         V27         V28         V29         V3
#>  [1,]  0.06852537  0.21312647  0.15603042  0.09653810  0.09420715 -0.2792255
#>  [2,] -0.06192516 -0.07354471 -0.01599304 -0.04509522 -0.02999928 -0.1952851
#>  [3,]  0.09763249  0.23632008  0.13385016  0.12162307  0.13413023 -0.2352313
#>  [4,]  0.24070809  0.13900386  0.22014113  0.31861815  0.23097714 -0.2629394
#>  [5,]  0.07951905  0.14238014  0.03790759  0.16893714  0.09078360 -0.2559340
#>  [6,]  0.13142235  0.03503885  0.16149724  0.13542269 -0.01972005 -0.1824533
#>  [7,] -0.03460150  0.08230576  0.14390984  0.15656045  0.03775495 -0.2661402
#>  [8,]  0.19764653  0.02419139  0.02841728  0.07757176  0.19288068 -0.2890865
#>  [9,]  0.15552918  0.13899054  0.15695737  0.03752646  0.03840597 -0.3068025
#> [10,]  0.04223800  0.11096796  0.22122351  0.20400917  0.23165995 -0.2518742
#>                 V30           V31          V32          V33          V34
#>  [1,]  0.0878299627  0.0583990584  0.028330108  0.061984674  0.054389439
#>  [2,] -0.1203432059  0.0085237459 -0.140215402 -0.113456849 -0.089408528
#>  [3,]  0.0381803133  0.0123309830  0.032067835 -0.032216658 -0.076346512
#>  [4,]  0.1084129441  0.1059151116  0.150810977  0.019600863  0.154559196
#>  [5,]  0.1932915615  0.1099752391 -0.002590166  0.006056395  0.022647400
#>  [6,]  0.0141179156 -0.0374513216 -0.097324398 -0.149295296 -0.102863672
#>  [7,]  0.1220659241  0.0049264379  0.071543414  0.009368160  0.039451692
#>  [8,]  0.0124732276 -0.0034629195  0.016742663  0.026160632  0.006072433
#>  [9,]  0.0009516986  0.0618683401  0.034211016 -0.011978465  0.009823185
#> [10,]  0.1993706907  0.0006574019 -0.025497952  0.089897296  0.024671962
#>                V35          V36          V37         V38         V39         V4
#>  [1,]  0.056955494  0.133292947 -0.053005692  0.09169883  0.09044752 -0.2455616
#>  [2,]  0.062534927 -0.140204551 -0.066437259 -0.08008524 -0.05325514 -0.1679460
#>  [3,] -0.055025749  0.001326949 -0.050061492 -0.01339075 -0.07911041 -0.2484518
#>  [4,]  0.102501281  0.045186975  0.039336347 -0.04345177 -0.07619516 -0.3494000
#>  [5,] -0.030479563  0.094402520 -0.094482598 -0.11842161  0.01365018 -0.2470600
#>  [6,] -0.103615064 -0.023785488 -0.203032498 -0.24330811 -0.26821525 -0.2192545
#>  [7,]  0.063944246 -0.028820724  0.014728553  0.04401239 -0.06537576 -0.2998540
#>  [8,]  0.009784783 -0.062564038  0.104637361  0.06110261 -0.02157371 -0.3138606
#>  [9,]  0.004321952 -0.078935073 -0.009031147 -0.12539387 -0.12318216 -0.3507843
#> [10,] -0.010220810  0.006751137 -0.026315930 -0.06947086  0.04496171 -0.1775500
#>               V40         V41          V42           V43          V44
#>  [1,] -0.08998851 -0.05310942 -0.142978946 -0.1390322773 -0.004194061
#>  [2,] -0.04458787 -0.17652464 -0.061525865 -0.1491919131 -0.201431503
#>  [3,] -0.02975466 -0.07914907 -0.014541830  0.0009705752 -0.098690550
#>  [4,]  0.04183276  0.03633545  0.001645269  0.0370416627 -0.127067115
#>  [5,]  0.05168448 -0.02328552 -0.063755935 -0.0244080145 -0.061459864
#>  [6,] -0.14058016 -0.14640394 -0.065904107 -0.1015985958 -0.165324836
#>  [7,] -0.10155574 -0.08747209 -0.125885500 -0.0887941829 -0.120502933
#>  [8,] -0.09640871  0.01726469 -0.020949091 -0.1495303334  0.006531162
#>  [9,] -0.03111086 -0.19405183 -0.135111481 -0.1454442985 -0.167694758
#> [10,] -0.15781292 -0.13914737 -0.160177344 -0.1457375658 -0.202192049
#>               V45         V46         V47        V48        V49         V5
#>  [1,] -0.22144533 -0.10439063 -0.25929980 -0.1766387 -0.2493499 -0.1517894
#>  [2,] -0.25422912 -0.16279291 -0.18534406 -0.2482747 -0.3108535 -0.3031884
#>  [3,] -0.04357582 -0.16949790 -0.09721049 -0.2575118 -0.3502376 -0.2257166
#>  [4,] -0.04038697 -0.05917328 -0.09676186 -0.2747111 -0.2690021 -0.1436655
#>  [5,] -0.08429917 -0.19969575 -0.14237883 -0.2765294 -0.1726916 -0.2500386
#>  [6,] -0.29481717 -0.16130949 -0.29409024 -0.1898367 -0.1610880 -0.1985894
#>  [7,] -0.10034015 -0.05989681 -0.21159606 -0.2303024 -0.2829400 -0.3017401
#>  [8,] -0.09959840 -0.17683904 -0.14555038 -0.1841315 -0.1791610 -0.2305341
#>  [9,] -0.10030373 -0.16992539 -0.22624316 -0.1702049 -0.2468381 -0.3605415
#> [10,] -0.21896376 -0.24804131 -0.20515880 -0.2805481 -0.3172738 -0.1680483
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3596999 -0.4001966 -0.3754167 -0.2834750 -0.3136518 -0.3526640
#>  [2,] -0.3044738 -0.2590204 -0.3493628 -0.2637435 -0.3221771 -0.2624675
#>  [3,] -0.3630260 -0.3125823 -0.3126891 -0.2346062 -0.4053040 -0.3879962
#>  [4,] -0.2503330 -0.3131775 -0.3228891 -0.3499115 -0.2346050 -0.2321005
#>  [5,] -0.2992727 -0.2922018 -0.3090848 -0.4187611 -0.3732574 -0.4123997
#>  [6,] -0.3081246 -0.2245915 -0.3813989 -0.2035887 -0.2403851 -0.3080046
#>  [7,] -0.3601450 -0.2154820 -0.2064995 -0.3928837 -0.3220341 -0.3323315
#>  [8,] -0.2970893 -0.3650250 -0.3716557 -0.3669463 -0.3425783 -0.3742283
#>  [9,] -0.3043868 -0.4324419 -0.4218838 -0.4452712 -0.3184793 -0.4528537
#> [10,] -0.2472401 -0.2841476 -0.2395179 -0.2362559 -0.3754347 -0.2626008
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3700447 -0.3267580 -0.3620472 -0.3848186 -0.25456790 -0.2914140
#>  [2,] -0.2327217 -0.2871090 -0.2058318 -0.2955744 -0.19662247 -0.3807483
#>  [3,] -0.3989448 -0.4329668 -0.3593133 -0.3014453 -0.08441352 -0.4315302
#>  [4,] -0.3481383 -0.2594978 -0.3983247 -0.2516309 -0.15731305 -0.4145500
#>  [5,] -0.3631053 -0.3900610 -0.2605918 -0.3351295 -0.28922280 -0.2688933
#>  [6,] -0.3039726 -0.3034412 -0.2682843 -0.2650492 -0.17430145 -0.4069495
#>  [7,] -0.3347336 -0.2067831 -0.2685834 -0.3810522 -0.21686660 -0.2768113
#>  [8,] -0.2715901 -0.2362712 -0.2977394 -0.3489685 -0.17604015 -0.3195721
#>  [9,] -0.3987497 -0.4629192 -0.3617147 -0.3742698 -0.32597464 -0.3931861
#> [10,] -0.2183495 -0.3321620 -0.2778888 -0.2266172 -0.18815674 -0.2196368
#>                V7         V8          V9
#>  [1,] -0.12223538 -0.2289953 -0.21662297
#>  [2,] -0.23474676 -0.1587659 -0.11089901
#>  [3,] -0.25767978 -0.0847048 -0.13726327
#>  [4,] -0.22731155 -0.1752142 -0.08622564
#>  [5,] -0.26799863 -0.2577653 -0.14354667
#>  [6,] -0.11529664 -0.1270540 -0.22445207
#>  [7,] -0.10017705 -0.1143854 -0.13901368
#>  [8,] -0.09253069 -0.1409877 -0.10922201
#>  [9,] -0.20375349 -0.1502615 -0.08805518
#> [10,] -0.08058412 -0.1454352 -0.11023731
#> 
#> $W[[2]]
#>             [,1]        [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] -0.08614600 -0.01520802  0.1315617 0.09020527 -0.01172015 0.01282267
#> [2,]  0.01251595  0.05707155 -0.1107863 0.04721165  0.03983427 0.07401449
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.01841930 0.04753906 0.03008535 -0.03177288
#> [2,] -0.09716105 0.03375967 0.04424395  0.04282291
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.559742e-04  0.0013878111  0.0018657147  0.0020217741  0.0017759880
#>  [2,]  5.598683e-05  0.0002888794  0.0004168389  0.0004811154  0.0003891639
#>  [3,] -1.208393e-04 -0.0004438212 -0.0007440503 -0.0007968419 -0.0004335981
#>  [4,]  8.962418e-05  0.0005249292  0.0006459246  0.0006834501  0.0006918103
#>  [5,]  2.136382e-04  0.0011869654  0.0015858311  0.0017080683  0.0015165832
#>  [6,]  1.391391e-04  0.0008246040  0.0011568777  0.0012942550  0.0011268230
#>  [7,]  9.752665e-06  0.0001590172  0.0001066216  0.0001186809  0.0002705287
#>  [8,]  1.415332e-04  0.0008406391  0.0010677852  0.0011199346  0.0010331424
#>  [9,]  1.441263e-04  0.0008315019  0.0010818136  0.0011753490  0.0010960062
#> [10,]  2.430866e-04  0.0013268451  0.0017730445  0.0019048047  0.0016806054
#>                 V14          V15          V16          V17          V18
#>  [1,]  0.0015415395 1.215906e-03 0.0011263334 0.0011812084 0.0013616606
#>  [2,]  0.0003147927 1.998863e-04 0.0001758203 0.0002070035 0.0002744665
#>  [3,] -0.0002806362 6.118986e-05 0.0003297924 0.0004918415 0.0005065070
#>  [4,]  0.0006691567 6.802426e-04 0.0007478142 0.0008296292 0.0008847626
#>  [5,]  0.0013174057 1.081339e-03 0.0010257197 0.0010837314 0.0012364649
#>  [6,]  0.0009230782 6.999886e-04 0.0006364293 0.0007081349 0.0008678496
#>  [7,]  0.0003288080 4.637474e-04 0.0006287226 0.0007462336 0.0007750183
#>  [8,]  0.0009250120 8.321949e-04 0.0008409853 0.0008951767 0.0009514492
#>  [9,]  0.0009761642 8.453673e-04 0.0008567280 0.0009329696 0.0010356586
#> [10,]  0.0014560497 1.177078e-03 0.0011038286 0.0011555852 0.0012978554
#>                V19            V2           V20           V21           V22
#>  [1,] 0.0020898955  3.186372e-04  0.0030347051  0.0034172482  0.0033732784
#>  [2,] 0.0004248162  6.810246e-05  0.0006447908  0.0007281677  0.0007143041
#>  [3,] 0.0001495358 -1.372837e-04 -0.0006035050 -0.0007646242 -0.0007507324
#>  [4,] 0.0010860207  1.165225e-04  0.0012696996  0.0013673147  0.0012962662
#>  [5,] 0.0018614903  2.686186e-04  0.0026145890  0.0029118761  0.0028455898
#>  [6,] 0.0013312028  1.814874e-04  0.0018409839  0.0020773875  0.0020755179
#>  [7,] 0.0007282351  1.881400e-05  0.0005535137  0.0005523473  0.0005358741
#>  [8,] 0.0013248655  1.844973e-04  0.0017282492  0.0019094801  0.0018272097
#>  [9,] 0.0014213900  1.833689e-04  0.0018710289  0.0020767651  0.0020352845
#> [10,] 0.0019797309  3.027017e-04  0.0028619588  0.0031971676  0.0031278139
#>                 V23           V24           V25          V26           V27
#>  [1,]  0.0033102924  0.0032943382  2.697797e-03 0.0023056562  0.0026533681
#>  [2,]  0.0007130314  0.0007086967  5.028215e-04 0.0003805088  0.0004862838
#>  [3,] -0.0007548364 -0.0005726001 -9.875366e-05 0.0001722343 -0.0001723415
#>  [4,]  0.0012042902  0.0012336188  1.134311e-03 0.0010405004  0.0010598652
#>  [5,]  0.0027385616  0.0027089382  2.240178e-03 0.0019092649  0.0021544871
#>  [6,]  0.0020090321  0.0019131705  1.453627e-03 0.0011608422  0.0013570843
#>  [7,]  0.0005138944  0.0006501135  8.095324e-04 0.0008754532  0.0007626581
#>  [8,]  0.0017074422  0.0017042318  1.512595e-03 0.0013648080  0.0014774086
#>  [9,]  0.0019729508  0.0020052577  1.745041e-03 0.0015551544  0.0017036201
#> [10,]  0.0030311365  0.0029983473  2.473723e-03 0.0021168998  0.0024178855
#>                 V28          V29            V3          V30          V31
#>  [1,]  2.696702e-03 0.0023000698  3.983453e-04 0.0020847790 0.0014455303
#>  [2,]  4.683818e-04 0.0003801672  9.368175e-05 0.0003702267 0.0002191397
#>  [3,] -7.358214e-05 0.0002866356 -2.107074e-04 0.0002225697 0.0004159069
#>  [4,]  1.119753e-03 0.0010716806  1.267542e-04 0.0009682950 0.0007501331
#>  [5,]  2.210930e-03 0.0018983302  3.269339e-04 0.0017090266 0.0011623765
#>  [6,]  1.361174e-03 0.0011136379  2.298065e-04 0.0009740626 0.0005840889
#>  [7,]  8.173118e-04 0.0009139089 -1.664005e-09 0.0008203793 0.0007623824
#>  [8,]  1.551423e-03 0.0013842833  2.083884e-04 0.0012409516 0.0009192054
#>  [9,]  1.761371e-03 0.0015817866  2.189997e-04 0.0014197159 0.0010354214
#> [10,]  2.478882e-03 0.0021186469  3.724114e-04 0.0019370632 0.0013365300
#>                V32          V33           V34          V35          V36
#>  [1,] 9.845078e-04 6.349136e-04  5.236992e-04 6.026539e-04 4.943123e-04
#>  [2,] 7.655138e-05 6.952302e-06 -2.400403e-05 4.785206e-05 6.533089e-05
#>  [3,] 6.134900e-04 9.295211e-04  1.004688e-03 8.797657e-04 1.067365e-03
#>  [4,] 6.294467e-04 5.811129e-04  5.782876e-04 5.623748e-04 5.373768e-04
#>  [5,] 8.142026e-04 5.314181e-04  4.482732e-04 4.893876e-04 3.965369e-04
#>  [6,] 3.492903e-04 1.078761e-04 -1.484260e-05 2.545874e-05 5.423980e-05
#>  [7,] 7.124382e-04 8.121994e-04  8.549920e-04 8.027719e-04 8.651529e-04
#>  [8,] 7.374710e-04 5.428160e-04  4.908233e-04 4.364524e-04 3.746681e-04
#>  [9,] 7.682202e-04 6.051764e-04  5.616471e-04 5.811441e-04 5.479893e-04
#> [10,] 9.218258e-04 5.940178e-04  5.128694e-04 5.545373e-04 4.213042e-04
#>                V37          V38          V39            V4          V40
#>  [1,] 4.484865e-04 0.0008598921 0.0012420217  4.964604e-04 0.0008325913
#>  [2,] 5.670789e-05 0.0001426305 0.0002438134  1.116146e-04 0.0001742862
#>  [3,] 9.935734e-04 0.0006186501 0.0003060126 -2.552345e-04 0.0005703965
#>  [4,] 5.035626e-04 0.0005705378 0.0006456853  1.643802e-04 0.0005313004
#>  [5,] 3.886119e-04 0.0007539951 0.0010767527  4.143445e-04 0.0007210558
#>  [6,] 1.446148e-04 0.0004550959 0.0007401394  2.867586e-04 0.0005025253
#>  [7,] 7.921368e-04 0.0006331416 0.0005439622  7.856476e-08 0.0006069116
#>  [8,] 4.211739e-04 0.0006209463 0.0008235990  2.685023e-04 0.0006164011
#>  [9,] 5.248434e-04 0.0007137205 0.0009048612  2.729210e-04 0.0006933278
#> [10,] 4.029556e-04 0.0008238163 0.0011873066  4.676787e-04 0.0007676550
#>                V41           V42           V43           V44           V45
#>  [1,] 0.0009218389  0.0014218640  0.0014711823  0.0011963984  0.0015242412
#>  [2,] 0.0001840569  0.0003118416  0.0003331957  0.0002556743  0.0003321398
#>  [3,] 0.0002554804 -0.0001767199 -0.0002525110 -0.0001482599 -0.0005696039
#>  [4,] 0.0004637706  0.0005529159  0.0005545015  0.0004766341  0.0005226510
#>  [5,] 0.0007755883  0.0011791460  0.0012301816  0.0010224952  0.0012957585
#>  [6,] 0.0005471754  0.0008628838  0.0008892690  0.0007180789  0.0009045980
#>  [7,] 0.0004255487  0.0002915917  0.0002550456  0.0002472167  0.0001140158
#>  [8,] 0.0006055050  0.0008029886  0.0008200814  0.0007176903  0.0008900430
#>  [9,] 0.0006731400  0.0008944816  0.0009060411  0.0007612200  0.0009137895
#> [10,] 0.0008429049  0.0013013174  0.0013635594  0.0011328112  0.0014760231
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.226027e-03  9.559519e-04  7.792677e-04  4.436191e-04  4.616881e-04
#>  [2,]  2.663606e-04  2.366290e-04  1.978249e-04  1.123352e-04  9.749851e-05
#>  [3,] -4.774939e-04 -4.066252e-04 -3.298816e-04 -1.913944e-04 -1.042307e-04
#>  [4,]  4.287578e-04  3.117699e-04  2.562815e-04  1.423481e-04  1.918023e-04
#>  [5,]  1.054182e-03  8.048917e-04  6.480416e-04  3.701337e-04  3.990553e-04
#>  [6,]  7.394638e-04  6.076805e-04  4.770787e-04  2.826031e-04  2.898482e-04
#>  [7,]  9.515984e-05  5.041739e-05  4.588152e-05  1.636877e-05  6.474082e-05
#>  [8,]  7.499877e-04  5.468226e-04  4.289422e-04  2.422850e-04  2.750919e-04
#>  [9,]  7.447518e-04  5.669786e-04  4.588571e-04  2.589893e-04  2.788747e-04
#> [10,]  1.206101e-03  9.144451e-04  7.332301e-04  4.174950e-04  4.331017e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.205674e-04  1.356770e-04  1.337609e-04  6.892924e-05  8.367554e-05
#>  [2,]  2.483989e-05  3.125693e-05  3.238893e-05  1.399880e-05  1.865861e-05
#>  [3,] -2.867965e-05 -4.972281e-05 -5.931810e-05 -1.448884e-05 -2.808379e-05
#>  [4,]  4.777657e-05  4.646630e-05  4.227981e-05  2.621892e-05  2.836191e-05
#>  [5,]  1.035081e-04  1.141975e-04  1.101678e-04  5.786074e-05  6.873244e-05
#>  [6,]  7.134364e-05  8.362248e-05  7.924579e-05  3.910590e-05  4.798461e-05
#>  [7,]  1.881575e-05  9.345647e-06  2.758642e-06  9.092846e-06  6.743290e-06
#>  [8,]  7.231311e-05  7.527774e-05  6.887420e-05  3.706707e-05  4.172101e-05
#>  [9,]  7.566990e-05  7.922609e-05  7.427026e-05  4.035583e-05  4.764721e-05
#> [10,]  1.170591e-04  1.271332e-04  1.233464e-04  6.370552e-05  7.615266e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  4.797160e-05  3.882278e-05  4.596732e-05  5.546422e-05  7.268749e-05
#>  [2,]  8.763649e-06  7.369231e-06  1.020210e-05  1.071612e-05  1.775417e-05
#>  [3,] -3.409825e-06 -2.175204e-06 -1.514874e-05 -1.350835e-05 -3.596379e-05
#>  [4,]  2.290522e-05  1.817843e-05  1.762123e-05  2.195923e-05  2.152462e-05
#>  [5,]  4.069892e-05  3.359977e-05  3.750999e-05  4.759568e-05  5.936633e-05
#>  [6,]  2.307516e-05  2.131820e-05  2.234804e-05  3.297585e-05  4.294320e-05
#>  [7,]  1.503258e-05  1.162924e-05  7.156013e-06  6.996415e-06 -4.208975e-07
#>  [8,]  2.920531e-05  2.386804e-05  2.418581e-05  3.099879e-05  3.471419e-05
#>  [9,]  3.219966e-05  2.569427e-05  2.729271e-05  3.290174e-05  3.943878e-05
#> [10,]  4.600987e-05  3.679684e-05  4.269315e-05  5.177675e-05  6.715182e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  6.151619e-04  4.587928e-05  6.517352e-04  0.0007110036  0.0011857636
#>  [2,]  1.487046e-04  9.850733e-06  1.627507e-04  0.0001570801  0.0002517285
#>  [3,] -1.697418e-04 -1.937723e-05 -1.417768e-04 -0.0001168353 -0.0003845513
#>  [4,]  2.229057e-04  1.592237e-05  2.435905e-04  0.0002876167  0.0004423836
#>  [5,]  5.058865e-04  3.832712e-05  5.399848e-04  0.0006047816  0.0010101161
#>  [6,]  3.849689e-04  2.528833e-05  4.181198e-04  0.0004410157  0.0006856855
#>  [7,]  7.104761e-05  2.795763e-06  9.201078e-05  0.0001327542  0.0001376546
#>  [8,]  3.289622e-04  2.441604e-05  3.501813e-04  0.0004221200  0.0007041363
#>  [9,]  3.601055e-04  2.622339e-05  3.897159e-04  0.0004405355  0.0007061526
#> [10,]  5.566286e-04  4.350721e-05  5.916051e-04  0.0006612018  0.0011329044
#> 
#> $vW[[2]]
#>             [,1]          [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.01990081  0.0007271933 -0.03714719 -0.03344493 -0.02554712 -0.01335185
#> [2,]  0.01990081 -0.0007271933  0.03714719  0.03344493  0.02554712  0.01335185
#>             [,7]        [,8]        [,9]      [,10]
#> [1,] -0.02264365 -0.03258718 -0.02284189 -0.0242663
#> [2,]  0.02264365  0.03258718  0.02284189  0.0242663
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.51797855 0.23494252 0.62993106 0.53281120 0.58100441 0.08019642
#>  [7] 0.48983394 0.61488549 0.60159911 0.50147682
#> 
#> $B[[2]]
#> [1]  0.09056503 -0.03862217
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0037647869 0.0007036196 0.0002202494 0.0017928774 0.0031471270
#>  [6] 0.0019792776 0.0013600645 0.0022170889 0.0025245499 0.0034841547
#> 
#> $vB[[2]]
#> [1] -0.04773769  0.04773769
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#>  [2,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#>  [3,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#>  [4,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [5,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#>  [6,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#>  [7,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#>  [8,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [9,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [10,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [11,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [12,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [13,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [14,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [15,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [16,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [17,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [18,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [19,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [20,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [21,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [22,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [23,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [24,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [25,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [26,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [27,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [28,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [29,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [30,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [31,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [32,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [33,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [34,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [35,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [36,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [37,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [38,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [39,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [40,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#>  [2,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#>  [3,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#>  [4,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [5,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#>  [6,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#>  [7,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#>  [8,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [9,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [10,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [11,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [12,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [13,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [14,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [15,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [16,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [17,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [18,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [19,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [20,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [21,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [22,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [23,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [24,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [25,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [26,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [27,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [28,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [29,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [30,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [31,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [32,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [33,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [34,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [35,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [36,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [37,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [38,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [39,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [40,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#>  [2,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#>  [3,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#>  [4,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [5,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#>  [6,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#>  [7,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#>  [8,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [9,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [10,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [11,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [12,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [13,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [14,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [15,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [16,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [17,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [18,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [19,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [20,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [21,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [22,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [23,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [24,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [25,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [26,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [27,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [28,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [29,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [30,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [31,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [32,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [33,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [34,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [35,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [36,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [37,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [38,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [39,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [40,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#>  [2,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#>  [3,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#>  [4,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [5,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#>  [6,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#>  [7,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#>  [8,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [9,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [10,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [11,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [12,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [13,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [14,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [15,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [16,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [17,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [18,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [19,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [20,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [21,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [22,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [23,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [24,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [25,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [26,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [27,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [28,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [29,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [30,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [31,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [32,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [33,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [34,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [35,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [36,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [37,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [38,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [39,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [40,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#>  [2,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#>  [3,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#>  [4,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [5,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#>  [6,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#>  [7,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#>  [8,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [9,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [10,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [11,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [12,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [13,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [14,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [15,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [16,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [17,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [18,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [19,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [20,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [21,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [22,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [23,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [24,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [25,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [26,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [27,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [28,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [29,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [30,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [31,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [32,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [33,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [34,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [35,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [36,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [37,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [38,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [39,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [40,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#>  [2,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#>  [3,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#>  [4,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [5,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#>  [6,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#>  [7,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#>  [8,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [9,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [10,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [11,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [12,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [13,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [14,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [15,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [16,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [17,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [18,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [19,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [20,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [21,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [22,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [23,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [24,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [25,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [26,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [27,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [28,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [29,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [30,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [31,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [32,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [33,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [34,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [35,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [36,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [37,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [38,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [39,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [40,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6505807 0.1540053 0.7405804 0.7773344 0.6030085 0.2507535 0.6263329
#>  [2,] 0.6499548 0.1903071 0.7877037 0.7825589 0.6188244 0.3317127 0.5983866
#>  [3,] 0.5988875 0.1808052 0.7386340 0.7804806 0.5966060 0.3388310 0.5430128
#>  [4,] 0.7015374 0.1966921 0.8699722 0.8433344 0.6934211 0.4894336 0.6790079
#>  [5,] 0.7097553 0.2241523 0.8160933 0.8241487 0.6780336 0.4360824 0.6615445
#>  [6,] 0.6374922 0.1737533 0.8502418 0.8007569 0.6460794 0.3976209 0.6719786
#>  [7,] 0.6546511 0.1895679 0.8557768 0.8207910 0.6686060 0.4287184 0.6947070
#>  [8,] 0.6251182 0.1603051 0.7340998 0.8133000 0.6152694 0.2741107 0.5759480
#>  [9,] 0.7492038 0.2785121 0.8309946 0.8497037 0.7431209 0.4862205 0.7213397
#> [10,] 0.6690037 0.2289744 0.8277815 0.7779522 0.6665185 0.4595353 0.6839424
#> [11,] 0.6062100 0.2441067 0.7795583 0.7335198 0.6563183 0.4050827 0.6223920
#> [12,] 0.6723924 0.2189501 0.8291130 0.7966266 0.6655383 0.4400380 0.6448754
#> [13,] 0.6272077 0.2005107 0.8341582 0.7861461 0.6373149 0.4107563 0.6680593
#> [14,] 0.6298025 0.1950882 0.7783447 0.7518996 0.5774812 0.3082255 0.6052682
#> [15,] 0.6556452 0.1963277 0.8498756 0.8205149 0.6678234 0.4625191 0.6664155
#> [16,] 0.7419630 0.2503902 0.8127452 0.8746974 0.7474468 0.4316930 0.6924658
#> [17,] 0.6756446 0.1901494 0.8621922 0.8047978 0.6719601 0.4674187 0.6648739
#> [18,] 0.6654085 0.1627385 0.8228601 0.8086688 0.6377718 0.3462623 0.6438931
#> [19,] 0.6656935 0.1854784 0.8583968 0.8083843 0.6660464 0.4587439 0.6744307
#> [20,] 0.7830843 0.2236440 0.8059211 0.9053924 0.7755530 0.4075188 0.7118968
#> [21,] 0.6800352 0.2491391 0.7890372 0.8450147 0.7099846 0.4188329 0.6667525
#> [22,] 0.7420227 0.2595780 0.8341854 0.8631888 0.7448718 0.4839063 0.6985310
#> [23,] 0.6092238 0.1419548 0.6779616 0.7778008 0.6181151 0.1875604 0.5232949
#> [24,] 0.6968103 0.1634265 0.8154589 0.8439248 0.6721953 0.3529870 0.6525460
#> [25,] 0.6414259 0.1477247 0.7388110 0.8316141 0.6511725 0.2549263 0.5893197
#> [26,] 0.6081015 0.1905263 0.6765265 0.7554093 0.5802765 0.2396443 0.5532275
#> [27,] 0.6676765 0.2977824 0.8016438 0.7917572 0.6645324 0.4550436 0.6751590
#> [28,] 0.7545765 0.2441458 0.8452543 0.8743622 0.7532170 0.4864173 0.7061635
#> [29,] 0.5342490 0.1594273 0.5595462 0.7116575 0.5285500 0.1275891 0.4923790
#> [30,] 0.6087240 0.1820733 0.7592111 0.7785874 0.6122699 0.2909747 0.5874431
#> [31,] 0.6759507 0.1798721 0.8217699 0.7971376 0.6454663 0.3728239 0.6303204
#> [32,] 0.6786829 0.1791790 0.8381236 0.8382467 0.6869051 0.3829587 0.6669245
#> [33,] 0.6251500 0.2128612 0.6799607 0.7883307 0.6613827 0.2531375 0.5917692
#> [34,] 0.7157109 0.1942198 0.8519417 0.8393191 0.6961123 0.4846530 0.6610250
#> [35,] 0.7461461 0.2449948 0.8735083 0.8430849 0.7395726 0.5608565 0.7075469
#> [36,] 0.5693024 0.2225449 0.8002762 0.7189483 0.5825027 0.3101413 0.6454483
#> [37,] 0.7861452 0.2573329 0.8261291 0.8985050 0.7594934 0.4744176 0.6971766
#> [38,] 0.7388541 0.2243626 0.7765978 0.8461197 0.7062335 0.3425417 0.6732131
#> [39,] 0.7325168 0.2856349 0.8090005 0.8422863 0.7141967 0.4855137 0.6661987
#> [40,] 0.7012066 0.2079490 0.8196295 0.8637332 0.7385495 0.4136013 0.6701397
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7620066 0.5614328 0.6370423
#>  [2,] 0.7529379 0.5897955 0.6195761
#>  [3,] 0.6983982 0.5690815 0.6348412
#>  [4,] 0.8293217 0.7062450 0.7496017
#>  [5,] 0.7859235 0.7184450 0.6818132
#>  [6,] 0.8293546 0.6404648 0.7021429
#>  [7,] 0.8403019 0.6583528 0.7330181
#>  [8,] 0.6855710 0.5463571 0.5947809
#>  [9,] 0.8082061 0.7215437 0.7704824
#> [10,] 0.8274321 0.6699086 0.7189271
#> [11,] 0.7149919 0.5793777 0.6440503
#> [12,] 0.7758468 0.6689378 0.6966160
#> [13,] 0.8251676 0.6226800 0.6910890
#> [14,] 0.7850757 0.5666823 0.6078677
#> [15,] 0.7999567 0.6669169 0.7267128
#> [16,] 0.7685305 0.6926686 0.7829362
#> [17,] 0.8283065 0.6763576 0.7212622
#> [18,] 0.8202989 0.6239768 0.6532650
#> [19,] 0.8312692 0.6621429 0.7151714
#> [20,] 0.7807638 0.7258113 0.7722653
#> [21,] 0.7475309 0.6477022 0.7290638
#> [22,] 0.7903407 0.7170922 0.7649956
#> [23,] 0.6808881 0.4520683 0.5177203
#> [24,] 0.7994748 0.6487316 0.6671371
#> [25,] 0.7031999 0.5482272 0.6044720
#> [26,] 0.6851589 0.4787285 0.5610277
#> [27,] 0.7894836 0.6423030 0.6951816
#> [28,] 0.8006609 0.7275095 0.7829574
#> [29,] 0.6204225 0.3450656 0.4548501
#> [30,] 0.7330205 0.5861652 0.6129126
#> [31,] 0.8103504 0.6441430 0.6584788
#> [32,] 0.7826861 0.6739276 0.6927981
#> [33,] 0.6804593 0.5301683 0.6157833
#> [34,] 0.8198017 0.7081627 0.7240034
#> [35,] 0.8406863 0.7674326 0.7778672
#> [36,] 0.7842040 0.5216470 0.6469102
#> [37,] 0.7763585 0.7436231 0.7927464
#> [38,] 0.7662398 0.6430753 0.7039789
#> [39,] 0.7743753 0.7226948 0.7338366
#> [40,] 0.7757180 0.6690736 0.7425112
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4562177 0.5437823
#>  [2,] 0.4553632 0.5446368
#>  [3,] 0.4557384 0.5442616
#>  [4,] 0.4452276 0.5547724
#>  [5,] 0.4461794 0.5538206
#>  [6,] 0.4528145 0.5471855
#>  [7,] 0.4496537 0.5503463
#>  [8,] 0.4577483 0.5422517
#>  [9,] 0.4390855 0.5609145
#> [10,] 0.4472770 0.5527230
#> [11,] 0.4534085 0.5465915
#> [12,] 0.4487563 0.5512437
#> [13,] 0.4526731 0.5473269
#> [14,] 0.4576763 0.5423237
#> [15,] 0.4488864 0.5511136
#> [16,] 0.4403042 0.5596958
#> [17,] 0.4483561 0.5516439
#> [18,] 0.4538111 0.5461889
#> [19,] 0.4493759 0.5506241
#> [20,] 0.4385581 0.5614419
#> [21,] 0.4452158 0.5547842
#> [22,] 0.4399896 0.5600104
#> [23,] 0.4616825 0.5383175
#> [24,] 0.4507301 0.5492699
#> [25,] 0.4565631 0.5434369
#> [26,] 0.4592403 0.5407597
#> [27,] 0.4466390 0.5533610
#> [28,] 0.4391916 0.5608084
#> [29,] 0.4670319 0.5329681
#> [30,] 0.4572357 0.5427643
#> [31,] 0.4519536 0.5480464
#> [32,] 0.4502031 0.5497969
#> [33,] 0.4537936 0.5462064
#> [34,] 0.4448194 0.5551806
#> [35,] 0.4386221 0.5613779
#> [36,] 0.4600349 0.5399651
#> [37,] 0.4367530 0.5632470
#> [38,] 0.4455133 0.5544867
#> [39,] 0.4404789 0.5595211
#> [40,] 0.4447064 0.5552936
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]        [,6]
#>  [1,] 0.6215927 -1.7035261 1.0489873 1.2501997 0.4180165 -1.09459766
#>  [2,] 0.6188405 -1.4480160 1.3111393 1.2806409 0.4845612 -0.70044899
#>  [3,] 0.4008317 -1.5109015 1.0388809 1.2684693 0.3913430 -0.66850796
#>  [4,] 0.8546294 -1.4070988 1.9007128 1.6832497 0.8161621 -0.04227205
#>  [5,] 0.8941958 -1.2416307 1.4901001 1.5447125 0.7447498 -0.25707702
#>  [6,] 0.5644959 -1.5592570 1.7364987 1.3910315 0.6018496 -0.41538776
#>  [7,] 0.6395471 -1.4528204 1.7806477 1.5217161 0.7018867 -0.28708190
#>  [8,] 0.5113301 -1.6559600 1.0155240 1.4715969 0.4695169 -0.97386525
#>  [9,] 1.0943702 -0.9518539 1.5926930 1.7322792 1.0622530 -0.05513214
#> [10,] 0.7036826 -1.2141114 1.5699854 1.2537727 0.6924806 -0.16221363
#> [11,] 0.4314086 -1.1302948 1.2630943 1.0125549 0.6469301 -0.38433113
#> [12,] 0.7190256 -1.2717953 1.5793540 1.3653421 0.6880737 -0.24100777
#> [13,] 0.5202565 -1.3831057 1.6153885 1.3018496 0.5637290 -0.36083964
#> [14,] 0.5313697 -1.4172811 1.2560460 1.1087692 0.3124419 -0.80842837
#> [15,] 0.6439471 -1.4094063 1.7336255 1.5198396 0.6983570 -0.15020524
#> [16,] 1.0561964 -1.0965323 1.4679473 1.9431465 1.0850412 -0.27494702
#> [17,] 0.7338274 -1.4490399 1.8336183 1.4165553 0.7170638 -0.13050995
#> [18,] 0.6874908 -1.6379915 1.5358461 1.4413834 0.5657063 -0.63550949
#> [19,] 0.6887710 -1.4796624 1.8020377 1.4395455 0.6903572 -0.16540059
#> [20,] 1.2837317 -1.2445556 1.4237214 2.2586306 1.2399369 -0.37423200
#> [21,] 0.7539336 -1.1032090 1.3191315 1.6960240 0.8953093 -0.32756643
#> [22,] 1.0565081 -1.0481633 1.6155851 1.8420317 1.0714458 -0.06439721
#> [23,] 0.4440507 -1.7991483 0.7444200 1.2528959 0.4815555 -1.46594063
#> [24,] 0.8321544 -1.6329508 1.4858786 1.6877257 0.7181311 -0.60593507
#> [25,] 0.5815584 -1.7525589 1.0397979 1.5971100 0.6241971 -1.07250878
#> [26,] 0.4393388 -1.4465943 0.7378546 1.1276734 0.3239086 -1.15463042
#> [27,] 0.6976950 -0.8578804 1.3966002 1.3355503 0.6835581 -0.18031273
#> [28,] 1.1231715 -1.1300831 1.6978545 1.9400917 1.1158440 -0.05434397
#> [29,] 0.1372110 -1.6624955 0.2393204 0.9034479 0.1143242 -1.92244584
#> [30,] 0.4419519 -1.5023635 1.1483595 1.2574533 0.4568637 -0.89065496
#> [31,] 0.7352245 -1.5172140 1.5283847 1.3684992 0.5991698 -0.52012138
#> [32,] 0.7477257 -1.5219196 1.6443322 1.6452398 0.7856900 -0.47700900
#> [33,] 0.5114657 -1.3077644 0.7535912 1.3148927 0.6694620 -1.08194839
#> [34,] 0.9232846 -1.4228204 1.7499121 1.6531704 0.8288528 -0.06140747
#> [35,] 1.0781624 -1.1254876 1.9323407 1.6813625 1.0437482  0.24463886
#> [36,] 0.2790054 -1.2508972 1.3880216 0.9392506 0.3330557 -0.79945902
#> [37,] 1.3018443 -1.0598773 1.5584381 2.1807228 1.1499043 -0.10241908
#> [38,] 1.0400213 -1.2404215 1.2459490 1.7044861 0.8771609 -0.65198790
#> [39,] 1.0074293 -0.9166795 1.4435286 1.6753389 0.9158544 -0.05796156
#> [40,] 0.8530502 -1.3373330 1.5138392 1.8466494 1.0384433 -0.34909754
#>              [,7]      [,8]        [,9]       [,10]
#>  [1,]  0.51651673 1.1637124  0.24697886  0.56254990
#>  [2,]  0.39874711 1.1143431  0.36311989  0.48774942
#>  [3,]  0.17247750 0.8396820  0.27810485  0.55304244
#>  [4,]  0.74921634 1.5808277  0.87721614  1.09648907
#>  [5,]  0.67018485 1.3005262  0.93676129  0.76211694
#>  [6,]  0.71714780 1.5810602  0.57738200  0.85752304
#>  [7,]  0.82221827 1.6604763  0.65596231  1.00998976
#>  [8,]  0.30616132 0.7794939  0.18596259  0.38376509
#>  [9,]  0.95111688 1.4383959  0.95213187  1.21103702
#> [10,]  0.77194915 1.5675363  0.70777169  0.93914599
#> [11,]  0.49971356 0.9197538  0.32021937  0.59298730
#> [12,]  0.59658837 1.2416253  0.70338502  0.83123496
#> [13,]  0.69942057 1.5517584  0.50093897  0.80521532
#> [14,]  0.42746519 1.2954941  0.26832765  0.43835781
#> [15,]  0.69201697 1.3860238  0.69427361  0.97800806
#> [16,]  0.81167234 1.2000319  0.81262492  1.28286016
#> [17,]  0.68509052 1.5736720  0.73708252  0.95073131
#> [18,]  0.59230164 1.5183740  0.50646220  0.63342193
#> [19,]  0.72829359 1.5946489  0.67285858  0.92063467
#> [20,]  0.90461437 1.2701230  0.97347346  1.22114686
#> [21,]  0.69353338 1.0854868  0.60895427  0.98987790
#> [22,]  0.84031217 1.3269803  0.93008351  1.18026561
#> [23,]  0.09324710 0.7578560 -0.19231730  0.07091107
#> [24,]  0.63024920 1.3830148  0.61346836  0.69526494
#> [25,]  0.36115390 0.8625822  0.19351031  0.42413387
#> [26,]  0.21371972 0.7775829 -0.08513747  0.24533384
#> [27,]  0.73161222 1.3218152  0.58537400  0.82445674
#> [28,]  0.87682311 1.3904302  0.98202307  1.28298468
#> [29,] -0.03048628 0.4913420 -0.64080038 -0.18109289
#> [30,]  0.35340516 1.0100020  0.34813486  0.45957199
#> [31,]  0.53359152 1.4522886  0.59339199  0.65652275
#> [32,]  0.69430763 1.2813888  0.72600342  0.81323322
#> [33,]  0.37128396 0.7558832  0.12081985  0.47168829
#> [34,]  0.66786523 1.5150044  0.88647722  0.96440730
#> [35,]  0.88349943 1.6633431  1.19387037  1.25328038
#> [36,]  0.59909092 1.2903356  0.08664212  0.60548508
#> [37,]  0.83388897 1.2445701  1.06488566  1.34156045
#> [38,]  0.72275373 1.1871993  0.58873692  0.86631764
#> [39,]  0.69104190 1.2331834  0.95786826  1.01417577
#> [40,]  0.70881714 1.2408846  0.70399823  1.05906183
#> 
#> $pre[[3]]
#>                [,1]      [,2]
#>  [1,]  0.0170850787 0.1926639
#>  [2,]  0.0229353856 0.2019592
#>  [3,]  0.0250147630 0.2025257
#>  [4,]  0.0171147574 0.2370871
#>  [5,]  0.0142267732 0.2303463
#>  [6,]  0.0242442623 0.2135498
#>  [7,]  0.0206291168 0.2226992
#>  [8,]  0.0238395245 0.1932504
#>  [9,]  0.0025927635 0.2474669
#> [10,]  0.0146461528 0.2263251
#> [11,]  0.0188066113 0.2057150
#> [12,]  0.0177406511 0.2234375
#> [13,]  0.0235082246 0.2133841
#> [14,]  0.0246192060 0.1943199
#> [15,]  0.0208381232 0.2260094
#> [16,]  0.0030164920 0.2429441
#> [17,]  0.0197817308 0.2270969
#> [18,]  0.0232814760 0.2085652
#> [19,]  0.0208650928 0.2240575
#> [20,] -0.0006705952 0.2463452
#> [21,]  0.0105167286 0.2305369
#> [22,]  0.0055758262 0.2467799
#> [23,]  0.0241428528 0.1777139
#> [24,]  0.0186926477 0.2164137
#> [25,]  0.0216337457 0.1958206
#> [26,]  0.0201355994 0.1835372
#> [27,]  0.0131066921 0.2273667
#> [28,]  0.0047447261 0.2491883
#> [29,]  0.0233893599 0.1554535
#> [30,]  0.0252759895 0.1967523
#> [31,]  0.0211196240 0.2139002
#> [32,]  0.0195434511 0.2193934
#> [33,]  0.0128185881 0.1981730
#> [34,]  0.0156639789 0.2372892
#> [35,]  0.0081039031 0.2548602
#> [36,]  0.0264725637 0.1866748
#> [37,]  0.0001169175 0.2544673
#> [38,]  0.0050930471 0.2239088
#> [39,]  0.0066603018 0.2458789
#> [40,]  0.0114406727 0.2335236
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4562177  0.4562177
#>  [2,] -0.4553632  0.4553632
#>  [3,]  0.5442616 -0.5442616
#>  [4,]  0.5547724 -0.5547724
#>  [5,]  0.5538206 -0.5538206
#>  [6,] -0.4528145  0.4528145
#>  [7,] -0.4496537  0.4496537
#>  [8,]  0.5422517 -0.5422517
#>  [9,] -0.4390855  0.4390855
#> [10,]  0.5527230 -0.5527230
#> [11,]  0.5465915 -0.5465915
#> [12,] -0.4487563  0.4487563
#> [13,] -0.4526731  0.4526731
#> [14,] -0.4576763  0.4576763
#> [15,]  0.5511136 -0.5511136
#> [16,] -0.4403042  0.4403042
#> [17,]  0.5516439 -0.5516439
#> [18,] -0.4538111  0.4538111
#> [19,]  0.5506241 -0.5506241
#> [20,] -0.4385581  0.4385581
#> [21,]  0.5547842 -0.5547842
#> [22,] -0.4399896  0.4399896
#> [23,] -0.4616825  0.4616825
#> [24,] -0.4507301  0.4507301
#> [25,]  0.5434369 -0.5434369
#> [26,]  0.5407597 -0.5407597
#> [27,] -0.4466390  0.4466390
#> [28,] -0.4391916  0.4391916
#> [29,] -0.4670319  0.4670319
#> [30,]  0.5427643 -0.5427643
#> [31,]  0.5480464 -0.5480464
#> [32,]  0.5497969 -0.5497969
#> [33,]  0.5462064 -0.5462064
#> [34,]  0.5551806 -0.5551806
#> [35,] -0.4386221  0.4386221
#> [36,] -0.4600349  0.4600349
#> [37,] -0.4367530  0.4367530
#> [38,]  0.5544867 -0.5544867
#> [39,] -0.4404789  0.4404789
#> [40,]  0.5552936 -0.5552936
#> 
#> $L
#> [1] 0.7000258 0.7237797 0.7991032 0.7499807 0.7139279 0.6915280
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3333333