Classification Linear Mixed Effect Learner

Generalized linear model with random effects. Calls lme4::glmer() from lme4.

Initial parameter values

family - Is set to stats::binomial(link = "logit").

Formula

Although most mlr3 learners don't allow to specify the formula manually, and automatically set it by valling task$formula(), this learner allows to set the formula because it's core functionality depends it. This means that it might not always use all features that are available in the task. Be aware, that this can sometimes lead to unexpected error messages, because mlr3 checks the compatibility between the learner and the task on all available features.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.glmer")
lrn("classif.glmer")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, lme4

Parameters

Id	Type	Default	Levels	Range

formula	untyped	-		-

start	untyped	NULL		-

verbose	integer	0		\([0, \infty)\)

offset	untyped	NULL		-

contrasts	untyped	NULL		-

optimizer	character	-	Nelder_Mead, bobyqa, nlminbwrap, nloptwrap	-

restart_edge	logical	FALSE	TRUE, FALSE	-

boundary.tol	numeric	1e-05		\([0, \infty)\)

calc.derivs	logical	TRUE	TRUE, FALSE	-

check.nobs.vs.rankZ	character	ignore	ignore, warning, message, stop	-

check.nobs.vs.nlev	character	stop	ignore, warning, message, stop	-

check.nlev.gtreq.5	character	ignore	ignore, warning, message, stop	-

check.nlev.gtr.1	character	stop	ignore, warning, message, stop	-

check.nobs.vs.nRE	character	stop	ignore, warning, message, stop	-

check.rankX	character	message+drop.cols	message+drop.cols, silent.drop.cols, warn+drop.cols, stop.deficient, ignore	-

check.scaleX	character	warning	warning, stop, silent.rescale, message+rescale, warn+rescale, ignore	-

check.formula.LHS	character	stop	ignore, warning, message, stop	-

family	untyped	"stats::binomial(link = \"logit\")"		-

nAGQ	integer	1		\([0, \infty)\)

mustart	untyped	-		-

etastart	untyped	-		-

check.conv.grad	untyped	"lme4::.makeCC(\"warning\", tol = 2e-3, relTol = NULL)"		-

check.conv.singular	untyped	"lme4::.makeCC( action = \"message\", tol = formals(lme4::isSingular)$tol)"		-

check.conv.hess	untyped	"lme4::.makeCC(action = \"warning\", tol = 1e-6)"		-

optCtrl	untyped	list()		-

tolPwrss	untyped	-		-

compDev	logical	TRUE	TRUE, FALSE	-

nAGQ0initStep	logical	TRUE	TRUE, FALSE	-

check.response.not.const	untyped	"stop"		-

newparams	untyped	NULL		-

re.form	untyped	NULL		-

random.only	logical	FALSE	TRUE, FALSE	-

allow.new.levels	logical	FALSE	TRUE, FALSE	-

na.action	untyped	"stats::na.pass"		-

References

Bates, M D (2010). “lme4: Mixed-effects modeling with R.”

See also

Author

sebffischer

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifGlmer

Methods

Public methods

LearnerClassifGlmer$new()

LearnerClassifGlmer$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifGlmer$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifGlmer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("classif.glmer")
print(learner)
#> <LearnerClassifGlmer:classif.glmer>: Generalized Linear Mixed Effect Regression
#> * Model: -
#> * Parameters: family=<family>
#> * Packages: mlr3, lme4
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: twoclass, weights

# available parameters:
learner$param_set$ids()
#>  [1] "formula"                  "start"                   
#>  [3] "verbose"                  "offset"                  
#>  [5] "contrasts"                "optimizer"               
#>  [7] "restart_edge"             "boundary.tol"            
#>  [9] "calc.derivs"              "check.nobs.vs.rankZ"     
#> [11] "check.nobs.vs.nlev"       "check.nlev.gtreq.5"      
#> [13] "check.nlev.gtr.1"         "check.nobs.vs.nRE"       
#> [15] "check.rankX"              "check.scaleX"            
#> [17] "check.formula.LHS"        "family"                  
#> [19] "nAGQ"                     "mustart"                 
#> [21] "etastart"                 "check.conv.grad"         
#> [23] "check.conv.singular"      "check.conv.hess"         
#> [25] "optCtrl"                  "tolPwrss"                
#> [27] "compDev"                  "nAGQ0initStep"           
#> [29] "check.response.not.const" "newparams"               
#> [31] "re.form"                  "random.only"             
#> [33] "allow.new.levels"         "na.action"