Classification Imbalanced Random Forest Src Learner

Imbalanced Random forest for classification between two classes. Calls randomForestSRC::imbalanced.rfsrc() from from randomForestSRC.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.imbalanced_rfsrc")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, randomForestSRC

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

method	character	rfq	rfq, brf, standard	-

block.size	integer	10		\([1, \infty)\)

fast	logical	FALSE	TRUE, FALSE	-

ratio	numeric	-		\([0, 1]\)

mtry	integer	-		\([1, \infty)\)

mtry.ratio	numeric	-		\([0, 1]\)

nodesize	integer	15		\([1, \infty)\)

nodedepth	integer	-		\([1, \infty)\)

splitrule	character	gini	gini, auc, entropy	-

nsplit	integer	10		\([0, \infty)\)

importance	character	FALSE	FALSE, TRUE, none, permute, random, anti	-

bootstrap	character	by.root	by.root, by.node, none, by.user	-

samptype	character	swor	swor, swr	-

samp	untyped	-		-

membership	logical	FALSE	TRUE, FALSE	-

sampsize	untyped	-		-

sampsize.ratio	numeric	-		\([0, 1]\)

na.action	character	na.omit	na.omit, na.impute	-

nimpute	integer	1		\([1, \infty)\)

ntime	integer	-		\([1, \infty)\)

cause	integer	-		\([1, \infty)\)

proximity	character	FALSE	FALSE, TRUE, inbag, oob, all	-

distance	character	FALSE	FALSE, TRUE, inbag, oob, all	-

forest.wt	character	FALSE	FALSE, TRUE, inbag, oob, all	-

xvar.wt	untyped	-		-

split.wt	untyped	-		-

forest	logical	TRUE	TRUE, FALSE	-

var.used	character	FALSE	FALSE, all.trees, by.tree	-

split.depth	character	FALSE	FALSE, all.trees, by.tree	-

seed	integer	-		\((-\infty, -1]\)

do.trace	logical	FALSE	TRUE, FALSE	-

statistics	logical	FALSE	TRUE, FALSE	-

get.tree	untyped	-		-

outcome	character	train	train, test	-

ptn.count	integer	0		\([0, \infty)\)

cores	integer	1		\([1, \infty)\)

save.memory	logical	FALSE	TRUE, FALSE	-

perf.type	character	-	gmean, misclass, brier, none	-

case.depth	logical	FALSE	TRUE, FALSE	-

Custom mlr3 parameters

mtry: This hyperparameter can alternatively be set via the added hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

sampsize: This hyperparameter can alternatively be set via the added hyperparameter sampsize.ratio as sampsize = max(ceiling(sampsize.ratio * n_obs), 1). Note that sampsize and sampsize.ratio are mutually exclusive.

cores: This value is set as the option rf.cores during training and is set to 1 by default.

References

O’Brien R, Ishwaran H (2019). “A random forests quantile classifier for class imbalanced data.” Pattern Recognition, 90, 232–249. doi:10.1016/j.patcog.2019.01.036 .

Chao C, Leo B (2004). “Using Random Forest to Learn Imbalanced Data.” University of California, Berkeley.

See also

Author

HarutyunyanLiana

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifImbalancedRandomForestSRC

Methods

Public methods

LearnerClassifImbalancedRandomForestSRC$new()

LearnerClassifImbalancedRandomForestSRC$importance()

LearnerClassifImbalancedRandomForestSRC$selected_features()

LearnerClassifImbalancedRandomForestSRC$oob_error()

LearnerClassifImbalancedRandomForestSRC$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifImbalancedRandomForestSRC$new()

Method importance()

The importance scores are extracted from the slot importance.

Usage

LearnerClassifImbalancedRandomForestSRC$importance()

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot var.used.

Usage

LearnerClassifImbalancedRandomForestSRC$selected_features()

Returns

character().

Method oob_error()

OOB error extracted from the model slot err.rate.

Usage

LearnerClassifImbalancedRandomForestSRC$oob_error()

Returns

numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifImbalancedRandomForestSRC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.imbalanced_rfsrc", importance = "TRUE")
print(learner)
#> <LearnerClassifImbalancedRandomForestSRC:classif.imbalanced_rfsrc>: Imbalanced Random Forest
#> * Model: -
#> * Parameters: importance=TRUE
#> * Packages: mlr3, randomForestSRC
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, oob_error, twoclass, weights

# Define a Task
task = mlr3::tsk("sonar")
# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#>                          Sample size: 139
#>            Frequency of class labels: 77, 62
#>                      Number of trees: 3000
#>            Forest terminal node size: 1
#>        Average no. of terminal nodes: 17.462
#> No. of variables tried at each split: 8
#>               Total no. of variables: 60
#>        Resampling used to grow trees: swor
#>     Resample size used to grow trees: 88
#>                             Analysis: RFQ
#>                               Family: class
#>                       Splitting rule: auc *random*
#>        Number of random split points: 10
#>                     Imbalanced ratio: 1.2419
#>                    (OOB) Brier score: 0.14845858
#>         (OOB) Normalized Brier score: 0.59383433
#>                            (OOB) AUC: 0.90155006
#>                       (OOB) Log-loss: 0.46077966
#>                         (OOB) PR-AUC: 0.87983077
#>                         (OOB) G-mean: 0.79469861
#>    (OOB) Requested performance error: 0.20530139
#> 
#> Confusion matrix:
#> 
#>           predicted
#>   observed  M  R class.error
#>          M 67 10      0.1299
#>          R 17 45      0.2742
#> 
#>       (OOB) Misclassification rate: 0.1942446
print(learner$importance())
#>           V12           V13           V48           V37           V43 
#>  0.0236797690  0.0178606764  0.0178606764  0.0147659603  0.0147659603 
#>           V44           V59           V20           V15           V19 
#>  0.0147659603  0.0119510357  0.0094128938  0.0088795927  0.0088795927 
#>            V2           V24           V25           V26           V27 
#>  0.0088795927  0.0088795927  0.0088795927  0.0088795927  0.0088795927 
#>           V28           V33           V39           V41           V42 
#>  0.0088795927  0.0088795927  0.0088795927  0.0088795927  0.0088795927 
#>           V49            V6           V60            V7           V32 
#>  0.0088795927  0.0088795927  0.0088795927  0.0088795927  0.0059528824 
#>           V34            V4           V16           V17           V18 
#>  0.0059528824  0.0059528824  0.0030369917  0.0030369917  0.0030369917 
#>           V29            V3           V45           V46           V47 
#>  0.0030369917  0.0030369917  0.0030369917  0.0030369917  0.0030369917 
#>           V21            V1           V14           V30           V36 
#>  0.0009230718  0.0000000000  0.0000000000  0.0000000000  0.0000000000 
#>           V40           V51           V53           V54           V58 
#>  0.0000000000  0.0000000000  0.0000000000  0.0000000000  0.0000000000 
#>            V8           V23           V22           V35           V38 
#>  0.0000000000 -0.0027628046 -0.0059086212 -0.0059086212 -0.0059086212 
#>            V5           V50           V52           V55           V57 
#> -0.0059086212 -0.0059086212 -0.0059086212 -0.0059086212 -0.0059086212 
#>           V10           V11           V56            V9           V31 
#> -0.0087814666 -0.0087814666 -0.0087814666 -0.0087814666 -0.0137196083 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#> 0.07246377