Classification Kernlab Support Vector Machine

Support vector machine for classification. Calls kernlab::ksvm() from kernlab.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.ksvm")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, kernlab

Parameters

Id	Type	Default	Levels	Range

scaled	logical	TRUE	TRUE, FALSE	-

type	character	C-svc	C-svc, nu-svc, C-bsvc, spoc-svc, kbb-svc	-

kernel	character	rbfdot	rbfdot, polydot, vanilladot, laplacedot, besseldot, anovadot	-

C	numeric	1		\((-\infty, \infty)\)

nu	numeric	0.2		\([0, \infty)\)

cache	integer	40		\([1, \infty)\)

tol	numeric	0.001		\([0, \infty)\)

shrinking	logical	TRUE	TRUE, FALSE	-

sigma	numeric	-		\([0, \infty)\)

degree	integer	-		\([1, \infty)\)

scale	numeric	-		\([0, \infty)\)

order	integer	-		\((-\infty, \infty)\)

offset	numeric	-		\((-\infty, \infty)\)

coupler	character	minpair	minpair, pkpd	-

References

Karatzoglou, Alexandros, Smola, Alex, Hornik, Kurt, Zeileis, Achim (2004). “kernlab-an S4 package for kernel methods in R.” Journal of statistical software, 11(9), 1–20.

See also

Author

mboecker

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifKSVM

Methods

Public methods

LearnerClassifKSVM$new()

LearnerClassifKSVM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifKSVM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifKSVM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.ksvm")
print(learner)
#> <LearnerClassifKSVM:classif.ksvm>: Support Vector Machine
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, kernlab
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, character, factor, ordered
#> * Properties: multiclass, twoclass, weights

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Support Vector Machine object of class "ksvm" 
#> 
#> SV type: C-svc  (classification) 
#>  parameter : cost C = 1 
#> 
#> Gaussian Radial Basis kernel function. 
#>  Hyperparameter : sigma =  0.010010618913507 
#> 
#> Number of Support Vectors : 98 
#> 
#> Objective Function Value : -56.6673 
#> Training error : 0.028777 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551