LiblineaR Classification Learner

L2 regularized support vector classification. Calls LiblineaR::LiblineaR() from LiblineaR.

Details

Type of SVC depends on type argument:

0 – L2-regularized logistic regression (primal)

1 - L2-regularized L2-loss support vector classification (dual)

3 - L2-regularized L1-loss support vector classification (dual)

2 – L2-regularized L2-loss support vector classification (primal)

4 – Support vector classification by Crammer and Singer

5 - L1-regularized L2-loss support vector classification

6 - L1-regularized logistic regression

7 - L2-regularized logistic regression (dual)

If number of records > number of features, type = 2 is faster than type = 1 (Hsu et al. 2003).

Note that probabilistic predictions are only available for types 0, 6, and 7. The default epsilon value depends on the type parameter, see LiblineaR::LiblineaR.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.liblinear")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “numeric”

Required Packages: mlr3, mlr3extralearners, LiblineaR

Parameters

Id	Type	Default	Levels	Range

type	integer	0		\([0, 7]\)

cost	numeric	1		\([0, \infty)\)

epsilon	numeric	-		\([0, \infty)\)

bias	numeric	1		\((-\infty, \infty)\)

cross	integer	0		\([0, \infty)\)

verbose	logical	FALSE	TRUE, FALSE	-

wi	untyped	NULL		-

findC	logical	FALSE	TRUE, FALSE	-

useInitC	logical	TRUE	TRUE, FALSE	-

References

Fan, Rong-En, Chang, Kai-Wei, Hsieh, Cho-Jui, Wang, Xiang-Rui, Lin, Chih-Jen (2008). “LIBLINEAR: A library for large linear classification.” the Journal of machine Learning research, 9, 1871–1874.

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLiblineaR

Methods

Public methods

LearnerClassifLiblineaR$new()

LearnerClassifLiblineaR$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLiblineaR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLiblineaR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.liblinear")
print(learner)
#> 
#> ── <LearnerClassifLiblineaR> (classif.liblinear): Support Vector Machine ───────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and LiblineaR
#> • Predict Types: [response] and prob
#> • Feature Types: numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $TypeDetail
#> [1] "L2-regularized logistic regression primal (L2R_LR)"
#> 
#> $Type
#> [1] 0
#> 
#> $W
#>              V1        V10      V11       V12        V13       V14      V15
#> [1,] -0.1844806 -0.8492289 -1.17067 -1.220684 -0.5026452 0.1892383 0.556138
#>            V16       V17       V18        V19         V2       V20        V21
#> [1,] 0.5427373 0.3464923 0.2505351 -0.2303755 -0.1828508 -0.032877 -0.3695701
#>             V22        V23         V24         V25       V26       V27
#> [1,] -0.6312237 -0.1364889 -0.06411637 -0.03298642 0.1858501 0.0193694
#>             V28         V29         V3        V30       V31        V32
#> [1,] -0.3640198 -0.04530168 -0.2089899 -0.0781761 0.5467645 -0.1066024
#>            V33       V34       V35      V36      V37         V38       V39
#> [1,] 0.2716804 0.8446323 0.4147343 1.190379 0.541686 -0.08938935 -0.144617
#>              V4       V40        V41        V42        V43        V44     V45
#> [1,] -0.4431606 0.3904124 -0.4041373 -0.3548932 -0.5615664 -0.9798186 -1.2676
#>             V46        V47        V48        V49        V5          V50
#> [1,] -0.9012961 -0.5445689 -0.5580557 -0.3577253 -0.192942 0.0006291315
#>              V51        V52         V53         V54         V55         V56
#> [1,] -0.06895403 -0.1074708 -0.03517974 -0.03054467 0.002486579 -0.02831278
#>              V57         V58         V59          V6         V60          V7
#> [1,] 0.008187519 -0.02683579 -0.03858701 -0.04335539 -0.01925416 0.006633323
#>            V8         V9      Bias
#> [1,] 0.171383 -0.7111422 0.7646374
#> 
#> $Bias
#> [1] 1
#> 
#> $ClassNames
#> [1] R M
#> Levels: M R
#> 
#> $NbClass
#> [1] 2
#> 
#> attr(,"class")
#> [1] "LiblineaR"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058