L1-Regularized L2-Loss Support Vector Classification Learner

Calls LiblineaR::LiblineaR from package LiblineaR.

Details

Calls LiblineaR::LiblineaR() with type = 5.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.liblinearl1l2svc")
lrn("classif.liblinearl1l2svc")

Traits

Packages: LiblineaR

Predict Types: response

Feature Types: numeric

Properties: multiclass, twoclass

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLiblineaRL1L2SVC

Methods

Public methods

LearnerClassifLiblineaRL1L2SVC$new()

LearnerClassifLiblineaRL1L2SVC$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLiblineaRL1L2SVC$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLiblineaRL1L2SVC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (mlr3misc::require_namespaces(c("LiblineaR"))) {
  learner = mlr3::lrn("classif.liblinearl1l2svc")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}

#> Error: The following packages could not be loaded: LiblineaR