Classification Light GBM Learner

Calls lightgbm::lgb.train from package lightgbm.

Details

For categorical features either pre-process data by encoding columns or specify the categorical columns with the categorical_feature parameter. For this learner please do not prefix the categorical feature with name:.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.lightgbm")
lrn("classif.lightgbm")

Meta Information

Task type: “classif”

Predict Types: “prob”, “response”

Feature Types: “numeric”, “integer”

Required Packages: mlr3extralearners, lightgbm

Parameters

Id	Type	Default	Levels	Range

nrounds	integer	5		\([1, \infty)\)

objective	character	binary	binary, multiclass, multiclassova, cross_entropy, cross_entropy_lambda, lambdarank, rank_xendcg	\((-\infty, \infty)\)

metric	character		, None, ndcg, map, auc, average_precision, binary_logloss, binary_error, auc_mu, multi_logloss, ...	\((-\infty, \infty)\)

custom_eval	list	NULL		\((-\infty, \infty)\)

verbose	integer	1		\((-\infty, \infty)\)

record	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

eval_freq	integer	1		\([1, \infty)\)

init_model	list	NULL		\((-\infty, \infty)\)

early_stopping_rounds	integer	-		\([1, \infty)\)

callbacks	list	-		\((-\infty, \infty)\)

reset_data	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

categorical_feature	list			\((-\infty, \infty)\)

boosting	character	gbdt	gbdt, rf, dart, goss	\((-\infty, \infty)\)

linear_tree	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_iterations	integer	100		\([0, \infty)\)

learning_rate	numeric	0.1		\([0, \infty)\)

num_leaves	integer	31		\([1, 131072]\)

tree_learner	character	serial	serial, feature, data, voting	\((-\infty, \infty)\)

num_threads	integer	0		\([0, \infty)\)

device_type	character	cpu	cpu, gpu	\((-\infty, \infty)\)

seed	integer	-		\((-\infty, \infty)\)

deterministic	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

force_col_wise	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

force_row_wise	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

histogram_pool_size	integer	-1		\((-\infty, \infty)\)

max_depth	integer	-1		\((-\infty, \infty)\)

min_data_in_leaf	integer	20		\([0, \infty)\)

min_sum_hessian_in_leaf	numeric	0.001		\([0, \infty)\)

bagging_fraction	numeric	1		\([0, 1]\)

pos_bagging_fraction	numeric	1		\([0, 1]\)

neg_bagging_fraction	numeric	1		\([0, 1]\)

bagging_freq	integer	0		\([0, \infty)\)

bagging_seed	integer	3		\((-\infty, \infty)\)

feature_fraction	numeric	1		\([0, 1]\)

feature_fraction_bynode	numeric	1		\([0, 1]\)

feature_fraction_seed	integer	2		\((-\infty, \infty)\)

extra_trees	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

extra_seed	integer	6		\((-\infty, \infty)\)

first_metric_only	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

max_delta_step	numeric	0		\((-\infty, \infty)\)

lambda_l1	numeric	0		\([0, \infty)\)

lambda_l2	numeric	0		\([0, \infty)\)

linear_lambda	numeric	0		\([0, \infty)\)

min_gain_to_split	numeric	0		\([0, \infty)\)

drop_rate	numeric	0.1		\([0, 1]\)

max_drop	integer	50		\((-\infty, \infty)\)

skip_drop	numeric	0.5		\([0, 1]\)

xgboost_dart_mode	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

uniform_drop	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

drop_seed	integer	4		\((-\infty, \infty)\)

top_rate	numeric	0.2		\([0, 1]\)

other_rate	numeric	0.1		\([0, 1]\)

min_data_per_group	integer	100		\([1, \infty)\)

max_cat_threshold	integer	32		\([1, \infty)\)

cat_l2	numeric	10		\([0, \infty)\)

cat_smooth	numeric	10		\([0, \infty)\)

max_cat_to_onehot	integer	4		\([1, \infty)\)

top_k	integer	20		\([1, \infty)\)

monotone_constraints	list	NULL		\((-\infty, \infty)\)

monotone_constraints_method	character	basic	basic, intermediate, advanced	\((-\infty, \infty)\)

monotone_penalty	numeric	0		\([0, \infty)\)

feature_contri	list	NULL		\((-\infty, \infty)\)

forcedsplits_filename	list			\((-\infty, \infty)\)

refit_decay_rate	numeric	0.9		\([0, 1]\)

cegb_tradeoff	numeric	1		\([0, \infty)\)

cegb_penalty_split	numeric	0		\([0, \infty)\)

cegb_penalty_feature_lazy	list	-		\((-\infty, \infty)\)

cegb_penalty_feature_coupled	list	-		\((-\infty, \infty)\)

path_smooth	numeric	0		\([0, \infty)\)

interaction_constraints	list	-		\((-\infty, \infty)\)

input_model	list			\((-\infty, \infty)\)

output_model	list	LightGBM_model.txt		\((-\infty, \infty)\)

saved_feature_importance_type	integer	0		\([0, 1]\)

snapshot_freq	integer	-1		\((-\infty, \infty)\)

max_bin	integer	255		\([2, \infty)\)

max_bin_by_feature	list	NULL		\((-\infty, \infty)\)

min_data_in_bin	integer	3		\([1, \infty)\)

bin_construct_sample_cnt	integer	200000		\([1, \infty)\)

data_random_seed	integer	1		\((-\infty, \infty)\)

is_enable_sparse	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

enable_bundle	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

use_missing	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

zero_as_missing	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

feature_pre_filter	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

pre_partition	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

two_round	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

header	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

group_column	list			\((-\infty, \infty)\)

forcedbins_filename	list			\((-\infty, \infty)\)

save_binary	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

objective_seed	integer	5		\((-\infty, \infty)\)

is_unbalance	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

scale_pos_weight	numeric	1		\([0, \infty)\)

sigmoid	numeric	1		\([0, \infty)\)

boost_from_average	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

lambdarank_truncation_level	integer	30		\([1, \infty)\)

lambdarank_norm	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

label_gain	list	-		\((-\infty, \infty)\)

metric_freq	integer	1		\([1, \infty)\)

is_provide_training_metric	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

eval_at	list	1, 2, 3, 4, 5		\((-\infty, \infty)\)

multi_error_top_k	integer	1		\([1, \infty)\)

auc_mu_weights	list	NULL		\((-\infty, \infty)\)

num_machines	integer	1		\([1, \infty)\)

local_listen_port	integer	12400		\([1, \infty)\)

time_out	integer	120		\([1, \infty)\)

machine_list_filename	list			\((-\infty, \infty)\)

machines	list			\((-\infty, \infty)\)

gpu_platform_id	integer	-1		\((-\infty, \infty)\)

gpu_device_id	integer	-1		\((-\infty, \infty)\)

gpu_use_dp	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_gpu	integer	1		\([1, \infty)\)

start_iteration	integer	0		\((-\infty, \infty)\)

num_iteration	integer	-1		\((-\infty, \infty)\)

pred_early_stop	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

pred_early_stop_freq	integer	10		\((-\infty, \infty)\)

pred_early_stop_margin	numeric	10		\((-\infty, \infty)\)

output_result	list	LightGBM_predict_result.txt		\((-\infty, \infty)\)

Custom mlr3 defaults

num_threads:

Actual default: 0L

Adjusted default: 1L

Reason for change: Prevents accidental conflicts with future.

verbose:

Actual default: 1L

Adjusted default: -1L

Reason for change: Prevents accidental conflicts with mlr messaging system.

See also

Author

kapsner

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLightGBM

Methods

Public methods

LearnerClassifLightGBM$new()

LearnerClassifLightGBM$importance()

LearnerClassifLightGBM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLightGBM$new()

Method importance()

The importance scores are extracted from lbg.importance.

Usage

LearnerClassifLightGBM$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLightGBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("lightgbm", quietly = TRUE)) {
  learner = mlr3::lrn("classif.lightgbm")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerClassifLightGBM:classif.lightgbm>
#> * Model: -
#> * Parameters: num_threads=1, verbose=-1
#> * Packages: mlr3, mlr3extralearners, lightgbm
#> * Predict Type: prob
#> * Feature types: numeric, integer
#> * Properties: importance, missings, multiclass, twoclass, weights
#>   [1] "nrounds"                       "objective"                    
#>   [3] "metric"                        "custom_eval"                  
#>   [5] "verbose"                       "record"                       
#>   [7] "eval_freq"                     "init_model"                   
#>   [9] "early_stopping_rounds"         "callbacks"                    
#>  [11] "reset_data"                    "categorical_feature"          
#>  [13] "boosting"                      "linear_tree"                  
#>  [15] "num_iterations"                "learning_rate"                
#>  [17] "num_leaves"                    "tree_learner"                 
#>  [19] "num_threads"                   "device_type"                  
#>  [21] "seed"                          "deterministic"                
#>  [23] "force_col_wise"                "force_row_wise"               
#>  [25] "histogram_pool_size"           "max_depth"                    
#>  [27] "min_data_in_leaf"              "min_sum_hessian_in_leaf"      
#>  [29] "bagging_fraction"              "pos_bagging_fraction"         
#>  [31] "neg_bagging_fraction"          "bagging_freq"                 
#>  [33] "bagging_seed"                  "feature_fraction"             
#>  [35] "feature_fraction_bynode"       "feature_fraction_seed"        
#>  [37] "extra_trees"                   "extra_seed"                   
#>  [39] "first_metric_only"             "max_delta_step"               
#>  [41] "lambda_l1"                     "lambda_l2"                    
#>  [43] "linear_lambda"                 "min_gain_to_split"            
#>  [45] "drop_rate"                     "max_drop"                     
#>  [47] "skip_drop"                     "xgboost_dart_mode"            
#>  [49] "uniform_drop"                  "drop_seed"                    
#>  [51] "top_rate"                      "other_rate"                   
#>  [53] "min_data_per_group"            "max_cat_threshold"            
#>  [55] "cat_l2"                        "cat_smooth"                   
#>  [57] "max_cat_to_onehot"             "top_k"                        
#>  [59] "monotone_constraints"          "monotone_constraints_method"  
#>  [61] "monotone_penalty"              "feature_contri"               
#>  [63] "forcedsplits_filename"         "refit_decay_rate"             
#>  [65] "cegb_tradeoff"                 "cegb_penalty_split"           
#>  [67] "cegb_penalty_feature_lazy"     "cegb_penalty_feature_coupled" 
#>  [69] "path_smooth"                   "interaction_constraints"      
#>  [71] "input_model"                   "output_model"                 
#>  [73] "saved_feature_importance_type" "snapshot_freq"                
#>  [75] "max_bin"                       "max_bin_by_feature"           
#>  [77] "min_data_in_bin"               "bin_construct_sample_cnt"     
#>  [79] "data_random_seed"              "is_enable_sparse"             
#>  [81] "enable_bundle"                 "use_missing"                  
#>  [83] "zero_as_missing"               "feature_pre_filter"           
#>  [85] "pre_partition"                 "two_round"                    
#>  [87] "header"                        "group_column"                 
#>  [89] "forcedbins_filename"           "save_binary"                  
#>  [91] "objective_seed"                "is_unbalance"                 
#>  [93] "scale_pos_weight"              "sigmoid"                      
#>  [95] "boost_from_average"            "lambdarank_truncation_level"  
#>  [97] "lambdarank_norm"               "label_gain"                   
#>  [99] "metric_freq"                   "is_provide_training_metric"   
#> [101] "eval_at"                       "multi_error_top_k"            
#> [103] "auc_mu_weights"                "num_machines"                 
#> [105] "local_listen_port"             "time_out"                     
#> [107] "machine_list_filename"         "machines"                     
#> [109] "gpu_platform_id"               "gpu_device_id"                
#> [111] "gpu_use_dp"                    "num_gpu"                      
#> [113] "start_iteration"               "num_iteration"                
#> [115] "pred_early_stop"               "pred_early_stop_freq"         
#> [117] "pred_early_stop_margin"        "output_result"