Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLogistic$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

LearnerClassifLogistic$unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

LearnerClassifLogistic$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                        -505.0415
#> V10                         141.528
#> V11                        129.2655
#> V12                         288.668
#> V13                       -275.8262
#> V14                        123.0802
#> V15                          36.941
#> V16                         -8.4414
#> V17                       -155.6309
#> V18                         73.9141
#> V19                        -52.9473
#> V2                         590.5522
#> V20                        174.0488
#> V21                        -91.3885
#> V22                         40.8339
#> V23                        -14.2732
#> V24                        -38.9948
#> V25                        143.3209
#> V26                       -250.5158
#> V27                        152.6273
#> V28                         22.6724
#> V29                        -96.4319
#> V3                        -993.2778
#> V30                        234.9589
#> V31                       -224.4029
#> V32                          16.636
#> V33                        121.3506
#> V34                       -216.9133
#> V35                        244.5779
#> V36                       -118.8021
#> V37                       -149.2613
#> V38                         97.7311
#> V39                         10.4629
#> V4                         210.4774
#> V40                       -129.6388
#> V41                         -5.4571
#> V42                         89.8016
#> V43                        135.5516
#> V44                       -138.4142
#> V45                         88.5193
#> V46                         16.7212
#> V47                         97.2387
#> V48                         50.3084
#> V49                        969.3567
#> V5                         263.9994
#> V50                       -2773.808
#> V51                        204.9281
#> V52                        -119.932
#> V53                       3021.6478
#> V54                       -758.9203
#> V55                       -1263.291
#> V56                        1624.703
#> V57                        -33.6535
#> V58                       1037.3341
#> V59                         450.034
#> V6                         398.7114
#> V60                       -673.2497
#> V7                        -402.8895
#> V8                        -170.4692
#> V9                          17.5489
#> Intercept                  -87.5234
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                                0
#> V10           2.9162506145045326E61
#> V11           1.3781390262959687E56
#> V12          2.3275636682550975E125
#> V13                               0
#> V14           2.8381775623476017E53
#> V15           1.1048071584294932E16
#> V16                          0.0002
#> V17                               0
#> V18           1.2603668530594381E32
#> V19                               0
#> V2           2.9754031430206712E256
#> V20            3.876503179958825E75
#> V21                               0
#> V22            5.419060901428663E17
#> V23                               0
#> V24                               0
#> V25           1.7517140177765782E62
#> V26                               0
#> V27           1.9283447678905915E66
#> V28                 7022751008.5817
#> V29                               0
#> V3                                0
#> V30          1.0999332449355223E102
#> V31                               0
#> V32                   16785576.5503
#> V33            5.033582119321696E52
#> V34                               0
#> V35           1.655130088922208E106
#> V36                               0
#> V37                               0
#> V38            2.780075921883554E42
#> V39                      34992.0267
#> V4            2.5654651703328235E91
#> V40                               0
#> V41                          0.0043
#> V42           1.0008143965082413E39
#> V43            7.401200558599137E58
#> V44                               0
#> V45           2.7762252277396705E38
#> V46                   18277381.2719
#> V47           1.6991747932355307E42
#> V48            7.057794402825152E21
#> V49                        Infinity
#> V5             4.50258196648045E114
#> V50                               0
#> V51            9.980704323339977E88
#> V52                               0
#> V53                        Infinity
#> V54                               0
#> V55                               0
#> V56                        Infinity
#> V57                               0
#> V58                        Infinity
#> V59          2.8006832993839306E195
#> V6            1.439386039053667E173
#> V60                               0
#> V7                                0
#> V8                                0
#> V9                    41820217.8439
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551