Classification Least Squares Support Vector Machine Learner

Least squares support vector machine for classification. Calls kernlab::lssvm() from kernlab. Parameters sigma, degree, scale, offset, order, length, lambda, and normalized are added to make tuning kpar easier. If kpar is provided then these new parameters are ignored. If none are provided then the default "automatic" is used for kpar.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lssvm")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, kernlab

Parameters

Id	Type	Default	Levels	Range

scaled	untyped	TRUE		-

kernel	character	rbfdot	rbfdot, polydot, vanilladot, tanhdot, laplacedot, besseldot, anovadot, splinedot, stringdot	-

sigma	numeric	-		\((-\infty, \infty)\)

degree	numeric	-		\((-\infty, \infty)\)

scale	numeric	-		\((-\infty, \infty)\)

offset	numeric	-		\((-\infty, \infty)\)

order	numeric	-		\((-\infty, \infty)\)

length	integer	-		\([0, \infty)\)

lambda	numeric	-		\((-\infty, \infty)\)

normalized	logical	-	TRUE, FALSE	-

kpar	untyped	"automatic"		-

tau	numeric	0.01		\((-\infty, \infty)\)

reduced	logical	TRUE	TRUE, FALSE	-

rank	integer	-		\((-\infty, \infty)\)

delta	integer	40		\((-\infty, \infty)\)

tol	numeric	1e-04		\((-\infty, \infty)\)

fit	logical	TRUE	TRUE, FALSE	-

na.action	untyped	na.omit		-

coupler	character	minpair	minpair, pkpd	-

References

Karatzoglou, Alexandros, Smola, Alex, Hornik, Kurt, Zeileis, Achim (2004). “kernlab-an S4 package for kernel methods in R.” Journal of statistical software, 11(9), 1–20.

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLSSVM

Methods

Public methods

LearnerClassifLSSVM$new()

LearnerClassifLSSVM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLSSVM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLSSVM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.lssvm")
print(learner)
#> <LearnerClassifLSSVM:classif.lssvm>: Support Vector Machine
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, kernlab
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric
#> * Properties: multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Using automatic sigma estimation (sigest) for RBF or laplace kernel 

print(learner$model)
#> Least Squares Support Vector Machine object of class "lssvm" 
#> 
#> problem type : classification 
#>  parameter : tau = 0.01 
#> 
#> Gaussian Radial Basis kernel function. 
#>  Hyperparameter : sigma =  0.0136771897783305 
#> 
#> Number of data points used for training : 46 
#> Training error : 0.043165 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696