Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.056346780 0.37434385 0.44687424 0.53726007 0.52253201 0.4069949
#>  [2,] 0.030361804 0.22337158 0.29673902 0.31843193 0.27206938 0.1963800
#>  [3,] 0.033037057 0.38185038 0.43078596 0.44877911 0.47979942 0.4217117
#>  [4,] 0.057328009 0.02744237 0.21051865 0.30903197 0.30888572 0.3864046
#>  [5,] 0.024071809 0.18699511 0.22212479 0.24715152 0.22330557 0.2926522
#>  [6,] 0.028788557 0.23743766 0.28990687 0.31123912 0.33374681 0.3531533
#>  [7,] 0.040959024 0.16924856 0.28144418 0.39093325 0.51634281 0.6185046
#>  [8,] 0.012243369 0.23253211 0.31191923 0.29616093 0.28157602 0.2410191
#>  [9,] 0.043705229 0.22830026 0.26477971 0.30917951 0.32176566 0.2669272
#> [10,] 0.032439484 0.19816592 0.20962375 0.18641811 0.22743644 0.3109720
#> [11,] 0.031020365 0.22683545 0.28139881 0.31252918 0.21674390 0.1511312
#> [12,] 0.043000000 0.71060000 0.73420000 0.50330000 0.30000000 0.1951000
#> [13,] 0.065943166 0.43650226 0.39017936 0.22768138 0.22648599 0.2734209
#> [14,] 0.014870784 0.13329861 0.07969621 0.02723422 0.03437061 0.1196941
#> [15,] 0.009154351 0.12457093 0.11333324 0.13256568 0.26152059 0.2152499
#> [16,] 0.011445188 0.10313968 0.08272975 0.17628994 0.04713319 0.2108162
#> [17,] 0.007829726 0.11140201 0.09531876 0.09640245 0.12688283 0.1084221
#> [18,] 0.028762358 0.08603048 0.08935527 0.22972779 0.19281816 0.1240143
#> [19,] 0.015979723 0.09859971 0.13440964 0.16464557 0.15907673 0.1652995
#> [20,] 0.009192354 0.11682936 0.14312041 0.06085535 0.04985114 0.1349816
#> [21,] 0.061575982 0.13989099 0.26891197 0.43868215 0.39098091 0.1901077
#> [22,] 0.006389452 0.04249788 0.18892434 0.39641930 0.26927399 0.2400894
#> [23,] 0.035800650 0.34662830 0.32207090 0.23298791 0.29640794 0.4936925
#> [24,] 0.016001696 0.18539119 0.17604075 0.19636809 0.35212573 0.4542939
#> [25,] 0.029300000 0.08600000 0.04140000 0.04720000 0.08350000 0.0938000
#>             V15        V16        V17       V18       V19          V2       V20
#>  [1,] 0.2925474 0.33949115 0.39361477 0.3928364 0.5261160 0.060788120 0.7046315
#>  [2,] 0.1613833 0.17416723 0.19158784 0.2283962 0.2460051 0.032812631 0.2464463
#>  [3,] 0.2804844 0.28637563 0.32609214 0.3439582 0.5028756 0.032242179 0.7017797
#>  [4,] 0.5219269 0.56138768 0.72382138 0.8595378 0.9547582 0.042871278 1.0289854
#>  [5,] 0.3428673 0.44062532 0.47945173 0.6325541 0.7441552 0.009009184 0.8635623
#>  [6,] 0.4218390 0.49375351 0.54988456 0.6340780 0.7375616 0.028095077 0.7962392
#>  [7,] 0.6675552 0.62508849 0.67993125 0.7692409 0.8740265 0.063477520 0.9056418
#>  [8,] 0.1980274 0.18264174 0.22293874 0.3528993 0.5920898 0.063201601 0.7071193
#>  [9,] 0.2058706 0.22383816 0.24538422 0.2097325 0.2539314 0.059022259 0.3255578
#> [10,] 0.4591598 0.59202337 0.68646536 0.7036277 0.7447331 0.045455572 0.8254799
#> [11,] 0.1143754 0.15598712 0.18187677 0.2927424 0.4739725 0.039746397 0.6076180
#> [12,] 0.2767000 0.37370000 0.25070000 0.2507000 0.3292000 0.090200000 0.4871000
#> [13,] 0.3445314 0.32765694 0.22982412 0.2670794 0.4691684 0.078723010 0.8162557
#> [14,] 0.0844794 0.07989574 0.20658746 0.2658831 0.2938263 0.011445098 0.4496565
#> [15,] 0.1670380 0.27377374 0.34809548 0.4406927 0.6635906 0.019323017 0.7623271
#> [16,] 0.1713851 0.22653561 0.09370399 0.2380670 0.4269985 0.021266751 0.3493607
#> [17,] 0.1462229 0.15950348 0.15503359 0.1949829 0.1750392 0.009794439 0.1752011
#> [18,] 0.1476954 0.19631632 0.18168600 0.1564514 0.1776549 0.042854260 0.1040220
#> [19,] 0.1657883 0.25945259 0.33348066 0.3740813 0.4227165 0.024005256 0.4418350
#> [20,] 0.2300679 0.17376649 0.13888781 0.1821498 0.2467201 0.020487609 0.2616830
#> [21,] 0.2608991 0.39321617 0.62896069 0.7430054 0.6411007 0.051057689 0.4406792
#> [22,] 0.1088171 0.04072918 0.11368006 0.3790803 0.2155990 0.061557600 0.5939579
#> [23,] 0.6252726 0.71012774 0.77752873 0.7163971 0.5340826 0.056517780 0.4559499
#> [24,] 0.5123516 0.38172322 0.17163583 0.2090622 0.2524757 0.036609997 0.3282449
#> [25,] 0.1466000 0.08090000 0.11790000 0.2179000 0.3326000 0.064400000 0.3258000
#>             V21        V22       V23       V24       V25       V26       V27
#>  [1,] 0.7574463 0.84886240 0.8689851 0.9220648 0.9927884 0.9597210 0.9549156
#>  [2,] 0.3538204 0.45760439 0.6024508 0.7194847 0.7757755 0.8313021 0.8988157
#>  [3,] 0.7385985 0.77346504 0.7450610 0.7478162 0.7662806 0.8790592 0.9495080
#>  [4,] 0.9831264 0.92005281 0.9107720 0.8591232 0.7101063 0.5091564 0.3659202
#>  [5,] 0.9239245 0.96194791 0.9407304 0.9633182 0.9949583 0.9287719 0.8433038
#>  [6,] 0.8534664 0.90546825 0.9494561 0.9758405 0.9087629 0.8919064 0.9221828
#>  [7,] 0.9628911 0.98986066 0.9275342 0.8956498 0.8269403 0.7583925 0.6890921
#>  [8,] 0.7897708 0.81115602 0.8200065 0.9249670 0.9382835 0.9633642 0.9957879
#>  [9,] 0.3488443 0.43084274 0.5621781 0.5937579 0.6198778 0.6560148 0.7280145
#> [10,] 0.8719731 0.85677351 0.7158209 0.5215973 0.3943000 0.2931867 0.2302354
#> [11,] 0.7531501 0.81022337 0.7824723 0.8781070 0.9142064 0.9757166 1.0048075
#> [12,] 0.6527000 0.84540000 0.9739000 1.0000000 0.6665000 0.5323000 0.4024000
#> [13,] 0.8296488 0.43062580 0.1550281 0.2215031 0.3449452 0.5832410 0.5630769
#> [14,] 0.5862015 0.72649463 0.8655218 0.8691411 0.8990450 0.9669500 0.9909461
#> [15,] 0.6805138 0.54574057 0.4988021 0.4858113 0.5642560 0.6699983 0.7802220
#> [16,] 0.3097844 0.40167327 0.5231322 0.5099784 0.3617577 0.5464830 0.5051029
#> [17,] 0.2575172 0.29158084 0.4352436 0.6579440 0.9133168 1.0080807 0.8404057
#> [18,] 0.1969897 0.05287756 0.2167910 0.2347721 0.0865086 0.2140337 0.4211563
#> [19,] 0.4004989 0.36613198 0.4067239 0.5691962 0.5815168 0.5563066 0.6128032
#> [20,] 0.3150947 0.54052349 0.6888746 0.7165055 0.7301607 0.7445464 0.6297538
#> [21,] 0.2532724 0.17259934 0.2187304 0.1675235 0.2507539 0.2564181 0.3000967
#> [22,] 0.7468839 0.75163467 0.7394784 0.8101324 1.0601749 1.0101904 0.7863926
#> [23,] 0.5220604 0.66358657 0.7312278 0.7340539 0.6497461 0.5601493 0.6367255
#> [24,] 0.3298731 0.31498888 0.5328166 0.7806188 0.7584242 0.7312986 0.6745627
#> [25,] 0.2111000 0.23020000 0.3361000 0.4259000 0.4609000 0.2606000 0.0874000
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.7045113 0.3893480 0.039708635 0.2972221 0.3408187 0.3866121 0.36661453
#>  [2,] 0.9772513 0.9132528 0.036927194 0.7125794 0.6372905 0.5315853 0.42576397
#>  [3,] 0.8559430 0.6597219 0.042951919 0.5381845 0.4869485 0.5805714 0.52637239
#>  [4,] 0.3335404 0.3845755 0.061574196 0.3393400 0.3096701 0.3441766 0.38084457
#>  [5,] 0.7072404 0.6492170 0.018140420 0.5703784 0.3980024 0.2543612 0.20831046
#>  [6,] 0.9233516 0.7641431 0.029480372 0.5758399 0.4471469 0.3390696 0.27093000
#>  [7,] 0.6527622 0.6389487 0.053760532 0.4381801 0.2691586 0.1588830 0.06226288
#>  [8,] 0.9701854 0.8631663 0.036095431 0.6387453 0.4153209 0.3323667 0.27576812
#>  [9,] 0.8142901 0.7697850 0.058045433 0.7999908 0.8517124 0.7779388 0.61613313
#> [10,] 0.2043940 0.3106459 0.038314107 0.4028414 0.2881454 0.3805375 0.38260633
#> [11,] 0.9539210 0.8345989 0.051066778 0.6538251 0.4949490 0.3663402 0.28976547
#> [12,] 0.3444000 0.4239000 0.083300000 0.4182000 0.4393000 0.1162000 0.43360000
#> [13,] 0.6545007 0.6581878 0.054753683 0.7255840 0.4388418 0.4474156 0.64339070
#> [14,] 0.8770869 0.7767902 0.014381709 0.6944263 0.6562760 0.5837479 0.53415275
#> [15,] 0.7946856 0.5946011 0.021983153 0.3688669 0.3558414 0.3717138 0.31841175
#> [16,] 0.2765488 0.3051101 0.056680830 0.4621995 0.6718183 0.6702146 0.55328523
#> [17,] 0.6686281 0.5500981 0.006481716 0.3732436 0.2871976 0.1748364 0.17160645
#> [18,] 0.6977619 0.7256342 0.027286382 0.6687347 0.6793555 0.5708384 0.67177943
#> [19,] 0.8093073 0.9417413 0.025110456 0.8548815 0.7697788 0.7005109 0.67350983
#> [20,] 0.7080413 0.7361042 0.025258541 0.6223179 0.4638424 0.4166849 0.64791951
#> [21,] 0.4626245 0.7224028 0.085500307 0.7514497 0.8412656 0.8404694 0.68502909
#> [22,] 0.6057789 0.5897043 0.074593114 0.5360694 0.5473758 0.4588150 0.58890617
#> [23,] 0.6230925 0.4543686 0.079231598 0.5075976 0.5048949 0.3512806 0.23229351
#> [24,] 0.8271366 0.9627055 0.041372844 0.8541172 0.7127683 0.5961507 0.59340043
#> [25,] 0.2862000 0.5606000 0.039000000 0.8344000 0.8096000 0.7250000 0.80480000
#>              V34        V35         V36        V37        V38         V39
#>  [1,] 0.31190444 0.11894937 0.365785652 0.44265739 0.33263503  0.19091690
#>  [2,] 0.26754286 0.14493919 0.098732609 0.18526211 0.15878314  0.18092695
#>  [3,] 0.36298615 0.14526040 0.246410365 0.30015420 0.36800388  0.48223512
#>  [4,] 0.47067645 0.46630966 0.411983466 0.42504991 0.41290607  0.44438218
#>  [5,] 0.13666316 0.19206737 0.205323103 0.24920935 0.30096991  0.37476822
#>  [6,] 0.11747293 0.08242893 0.081882753 0.14524240 0.22745611  0.21029325
#>  [7,] 0.03366307 0.12712463 0.243999353 0.25998327 0.31796623  0.40659597
#>  [8,] 0.24414304 0.24538206 0.161679143 0.11827377 0.11645771  0.16619226
#>  [9,] 0.60670519 0.61381217 0.515064827 0.39045220 0.40799767  0.42838389
#> [10,] 0.23301845 0.16228192 0.207516425 0.26864615 0.27226936  0.18831320
#> [11,] 0.19958904 0.11808542 0.003377035 0.06325476 0.12641354  0.09654152
#> [12,] 0.65530000 0.61720000 0.437300000 0.41180000 0.36410000  0.45720000
#> [13,] 0.68684207 0.71844378 0.630343230 0.59857630 0.64757448  0.59207195
#> [14,] 0.54247333 0.52892314 0.469778652 0.42442587 0.32234615  0.15392229
#> [15,] 0.27000746 0.29673953 0.322740515 0.38829124 0.35898049  0.31797564
#> [16,] 0.37873651 0.44834271 0.385031378 0.55050314 0.84209011  0.96554405
#> [17,] 0.19470559 0.26660189 0.219594400 0.10121968 0.09609672  0.14599739
#> [18,] 0.81896014 0.87920019 0.910385142 0.77566997 0.70677828  0.70042804
#> [19,] 0.68261693 0.63198708 0.603782545 0.58846665 0.41056118  0.26956324
#> [20,] 0.91324041 0.87886168 0.688931388 0.37303673 0.30952584  0.28402456
#> [21,] 0.69659101 0.73249280 0.814618870 0.89886304 1.01339620  0.88812437
#> [22,] 0.49568388 0.55409937 0.613497649 0.56069823 0.26050658 -0.10774038
#> [23,] 0.15403234 0.37652645 0.549511418 0.59144835 0.39366977  0.26845332
#> [24,] 0.52131594 0.41597854 0.338480543 0.19578843 0.19071218  0.27493198
#> [25,] 0.94350000 1.00000000 0.896000000 0.55160000 0.30370000  0.23380000
#>               V4       V40        V41        V42        V43          V44
#>  [1,] 0.02938183 0.1427838 0.12666932 0.25433812 0.42431190  0.453575838
#>  [2,] 0.05436521 0.1868252 0.25220126 0.26373199 0.20601570  0.123591335
#>  [3,] 0.03470709 0.3937921 0.24698386 0.37073840 0.39921476  0.447328301
#>  [4,] 0.05505464 0.3852984 0.31412902 0.32295048 0.29298801  0.263515068
#>  [5,] 0.08220613 0.3777201 0.34044393 0.39222312 0.30983352  0.183958376
#>  [6,] 0.07834002 0.2098293 0.25527508 0.25484256 0.20554348  0.169675282
#>  [7,] 0.03022446 0.2986583 0.16312661 0.21687994 0.19356570  0.198210981
#>  [8,] 0.02528952 0.1600157 0.06178622 0.05826809 0.15643932  0.175680422
#>  [9,] 0.05685140 0.4519496 0.57796915 0.57311985 0.48923316  0.450480462
#> [10,] 0.05223012 0.1062329 0.14385557 0.17412568 0.12052907  0.115719164
#> [11,] 0.04356088 0.1016949 0.02154571 0.11989992 0.20722675  0.214928123
#> [12,] 0.08130000 0.4367000 0.29640000 0.43120000 0.41550000  0.182400000
#> [13,] 0.06355747 0.4452085 0.34850682 0.31306162 0.27867872  0.252140781
#> [14,] 0.01637113 0.1515563 0.16340779 0.06276724 0.01811644 -0.002846467
#> [15,] 0.02390109 0.2499990 0.16137868 0.11510737 0.08311038  0.077161254
#> [16,] 0.02095464 0.9184426 0.65417129 0.55758052 0.42411568  0.421223188
#> [17,] 0.02308936 0.2086298 0.26593658 0.24590380 0.18058074  0.098870628
#> [18,] 0.02470584 0.7256315 0.65797062 0.57628139 0.41586769  0.233440387
#> [19,] 0.03394449 0.2195264 0.18881321 0.18054560 0.17718330  0.218410813
#> [20,] 0.03752422 0.2770895 0.22836256 0.30341200 0.37596436  0.264273710
#> [21,] 0.03143861 0.7282598 0.72445705 0.62714494 0.48246183  0.366974858
#> [22,] 0.06382765 0.2116570 0.29895007 0.57433006 0.49519228  0.282491607
#> [23,] 0.07006309 0.3739133 0.38967065 0.31538550 0.20772897  0.174766629
#> [24,] 0.03534048 0.2086438 0.15959790 0.18649605 0.22950339  0.173149308
#> [25,] 0.01730000 0.2382000 0.33180000 0.38210000 0.15750000  0.222800000
#>               V45        V46        V47        V48         V49         V5
#>  [1,] 0.323772468 0.15224767 0.05276821 0.04495715 0.030053845 0.02290524
#>  [2,] 0.111117962 0.15868259 0.17967143 0.14690540 0.071411117 0.07230373
#>  [3,] 0.462632398 0.27296984 0.15908628 0.14261435 0.074984771 0.08846645
#>  [4,] 0.144231986 0.07369676 0.05187805 0.01229466 0.029062907 0.11550791
#>  [5,] 0.133396319 0.13649058 0.12792400 0.09074222 0.065367547 0.10937166
#>  [6,] 0.139358356 0.06985026 0.06819978 0.06329572 0.035721543 0.14228761
#>  [7,] 0.132987825 0.11140937 0.11492083 0.10626295 0.077130031 0.10135185
#>  [8,] 0.149496836 0.16650096 0.11806286 0.06653284 0.032775637 0.01685783
#>  [9,] 0.536373738 0.48990456 0.32964049 0.20216517 0.110655096 0.06632967
#> [10,] 0.094979498 0.04102703 0.06188632 0.06041425 0.025659536 0.06740006
#> [11,] 0.192878310 0.20906410 0.17779341 0.11064125 0.055109870 0.06216935
#> [12,] 0.148700000 0.01380000 0.11640000 0.20520000 0.106900000 0.01650000
#> [13,] 0.330585516 0.25550243 0.13756800 0.10880556 0.060321227 0.09633857
#> [14,] 0.009097092 0.01701675 0.01553247 0.02522774 0.008282611 0.04069476
#> [15,] 0.058125043 0.04021439 0.04129542 0.04571398 0.019865103 0.01989782
#> [16,] 0.465358434 0.38316936 0.27577635 0.16313922 0.073351317 0.04480424
#> [17,] 0.092879504 0.10038212 0.04541815 0.03422480 0.027593805 0.04797034
#> [18,] 0.213446251 0.17380975 0.15839257 0.12622313 0.085808780 0.06688991
#> [19,] 0.204355968 0.15275323 0.08869793 0.06209519 0.034429245 0.04766453
#> [20,] 0.091294759 0.02603195 0.07292885 0.08235648 0.030237626 0.07491117
#> [21,] 0.107664687 0.09103426 0.12937885 0.12111380 0.057456548 0.04129858
#> [22,] 0.158799552 0.10339632 0.05983915 0.06930817 0.010651491 0.16710441
#> [23,] 0.175124081 0.11392494 0.12778841 0.11032701 0.066316447 0.07762114
#> [24,] 0.036156902 0.05565079 0.04969385 0.02944522 0.009820104 0.04730631
#> [25,] 0.158200000 0.14330000 0.16340000 0.11330000 0.056700000 0.04760000
#>                V50          V51          V52         V53         V54
#>  [1,]  0.008717839  0.013981052  0.005096078 0.003512286 0.008587241
#>  [2,]  0.017288474  0.019269616  0.015232423 0.008699139 0.010162125
#>  [3,]  0.024769930  0.022013361  0.014198989 0.009269609 0.010835215
#>  [4,]  0.009388816 -0.002132833  0.005107017 0.006641548 0.021996566
#>  [5,]  0.032245654  0.020784029  0.013360430 0.020725111 0.010670281
#>  [6,]  0.010694929  0.014184709  0.016825642 0.014282688 0.014156440
#>  [7,]  0.025930549  0.027859697  0.020396000 0.014915041 0.033505417
#>  [8,]  0.007370942  0.019987807  0.013286521 0.009492820 0.009881577
#>  [9,]  0.042900114  0.031677974  0.026292987 0.014265400 0.013653484
#> [10,]  0.016002171  0.016595242  0.012892313 0.008500051 0.007491519
#> [11,]  0.021307462  0.017076994  0.010550052 0.005778443 0.008010917
#> [12,]  0.019900000  0.020800000  0.017600000 0.019700000 0.021000000
#> [13,]  0.028287553  0.016658692  0.014885813 0.010554033 0.014370861
#> [14,]  0.004558895  0.001946323  0.003135731 0.008288861 0.004225234
#> [15,]  0.016437662  0.016755756  0.018185416 0.015351540 0.018039284
#> [16,]  0.027063444  0.023356787  0.011593348 0.006111989 0.013938921
#> [17,]  0.007779404  0.007407862  0.009559168 0.004575471 0.004889744
#> [18,]  0.030660375  0.013942057  0.009660519 0.016370139 0.014951636
#> [19,]  0.014284966  0.011368891  0.007281855 0.007489784 0.009216978
#> [20,]  0.018505752  0.027367544  0.010078186 0.012543727 0.013549379
#> [21,]  0.022681953  0.013010098  0.012741509 0.019610499 0.002967127
#> [22,] -0.002284750  0.003234198 -0.001675859 0.011920413 0.008839448
#> [23,]  0.016032518  0.016514469  0.012639207 0.009158525 0.010671666
#> [24,]  0.018721994  0.015802025  0.014066876 0.011850132 0.005960724
#> [25,]  0.013300000  0.017000000  0.003500000 0.005200000 0.008300000
#>               V55         V56          V57         V58         V59         V6
#>  [1,] 0.006424509 0.008920201 0.0076317277 0.007464799 0.004129138 0.12079094
#>  [2,] 0.007442844 0.007101897 0.0052660596 0.006839733 0.007487009 0.11851958
#>  [3,] 0.012416185 0.009904942 0.0077071352 0.008880904 0.008758967 0.11077909
#>  [4,] 0.021609584 0.016514017 0.0136060362 0.016179331 0.015990202 0.08684235
#>  [5,] 0.006156882 0.008797623 0.0075842026 0.005522407 0.006850081 0.12841535
#>  [6,] 0.002866250 0.011159753 0.0052883274 0.009721402 0.006618239 0.16404219
#>  [7,] 0.014918940 0.006323272 0.0058070468 0.009069952 0.011777363 0.17553743
#>  [8,] 0.004188709 0.002923582 0.0038560741 0.005257203 0.005741282 0.05498211
#>  [9,] 0.015535015 0.009320512 0.0084417018 0.009579133 0.012312887 0.13535298
#> [10,] 0.006763684 0.008270386 0.0074919082 0.006640282 0.005032866 0.09702093
#> [11,] 0.003021540 0.002094084 0.0027697484 0.001979730 0.002860744 0.07815741
#> [12,] 0.014100000 0.004900000 0.0027000000 0.016200000 0.005900000 0.02770000
#> [13,] 0.012461574 0.008467674 0.0090587338 0.011678283 0.011185868 0.09711840
#> [14,] 0.005333320 0.004653906 0.0024094946 0.004058838 0.007059344 0.07918301
#> [15,] 0.015775174 0.010390242 0.0103042909 0.014968657 0.005355072 0.04947452
#> [16,] 0.009355070 0.009972289 0.0041848206 0.007616281 0.008664213 0.05604251
#> [17,] 0.002713724 0.001584300 0.0002760923 0.002249367 0.001532624 0.07837941
#> [18,] 0.010387062 0.009074684 0.0098533342 0.008994985 0.008575706 0.10831366
#> [19,] 0.007647748 0.004896057 0.0072275954 0.005212921 0.008794263 0.09647701
#> [20,] 0.003800213 0.012381236 0.0157535730 0.006444556 0.007845010 0.08319444
#> [21,] 0.001983552 0.001246565 0.0012558658 0.016025950 0.010463758 0.06943903
#> [22,] 0.012928378 0.007949501 0.0093101409 0.006794649 0.001255219 0.21268525
#> [23,] 0.008880139 0.008195903 0.0129145852 0.009798355 0.010703151 0.11611505
#> [24,] 0.008239542 0.010208893 0.0048450542 0.004897229 0.013622201 0.11911723
#> [25,] 0.007800000 0.007500000 0.0105000000 0.016000000 0.009500000 0.08160000
#>               V60         V7         V8          V9
#>  [1,] 0.004926135 0.10537746 0.17807912  0.32245090
#>  [2,] 0.004203284 0.13237493 0.12921689  0.15741743
#>  [3,] 0.005186569 0.13332634 0.15586833  0.29448593
#>  [4,] 0.008768698 0.08179049 0.06056415  0.08737721
#>  [5,] 0.004648217 0.16711972 0.10473802  0.11503858
#>  [6,] 0.003775230 0.15211573 0.19182355  0.21635771
#>  [7,] 0.008279115 0.18030884 0.15710949  0.16658411
#>  [8,] 0.003471427 0.10433316 0.10625575  0.17426236
#>  [9,] 0.007827164 0.11942671 0.15571010  0.22381523
#> [10,] 0.004677252 0.11694186 0.10379230  0.16828917
#> [11,] 0.005319720 0.10185002 0.11529178  0.17927203
#> [12,] 0.002100000 0.05690000 0.20570000  0.38870000
#> [13,] 0.007553303 0.14063470 0.26134302  0.44258031
#> [14,] 0.005032977 0.08967581 0.07313047  0.13503592
#> [15,] 0.004317990 0.07280644 0.07583970  0.11415137
#> [16,] 0.013141388 0.11847842 0.14402992  0.05812555
#> [17,] 0.001886372 0.08541301 0.06000907  0.10571357
#> [18,] 0.008810764 0.12335146 0.10560989  0.02922177
#> [19,] 0.004923394 0.10139882 0.08330458  0.09342450
#> [20,] 0.010543056 0.11610862 0.10198375  0.10331040
#> [21,] 0.007312480 0.06173858 0.07919563  0.19373584
#> [22,] 0.017355179 0.15579472 0.13874411 -0.01169539
#> [23,] 0.008909587 0.15688203 0.22636824  0.28610135
#> [24,] 0.006768513 0.14529647 0.17295092  0.19428291
#> [25,] 0.001100000 0.09930000 0.03150000  0.07360000
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406