Classification L1 and L2 Penalized Learner

Penalized (L1 and L2) Logistic Regression. Calls penalized::penalized() from: penalized.

Details

The penalized and unpenalized arguments in the learner are implemented slightly differently than in penalized::penalized(). Here, there is no parameter for penalized but instead it is assumed that every variable is penalized unless stated in the unpenalized parameter.

Initial parameter values

trace is set to "FALSE" to disable printing output during model training.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.penalized")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, penalized

Parameters

Id	Type	Default	Levels	Range

epsilon	numeric	1e-10		\([0, \infty)\)

fusedl	logical	FALSE	TRUE, FALSE	-

lambda1	numeric	0		\([0, \infty)\)

lambda2	numeric	0		\([0, \infty)\)

maxiter	integer	-		\([1, \infty)\)

positive	untyped	FALSE		-

standardize	logical	FALSE	TRUE, FALSE	-

startbeta	untyped	-		-

startgamma	untyped	-		-

steps	untyped	1L		-

trace	logical	TRUE	TRUE, FALSE	-

unpenalized	untyped	-		-

References

Goeman, J J (2010). “L1 penalized estimation in the Cox proportional hazards model.” Biometrical journal, 52(1), 70–84.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifPenalized

Methods

Public methods

LearnerClassifPenalized$new()

LearnerClassifPenalized$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifPenalized$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifPenalized$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.penalized")
print(learner)
#> 
#> ── <LearnerClassifPenalized> (classif.penalized): Penalized Logistic Regression 
#> • Model: -
#> • Parameters: trace=FALSE
#> • Packages: mlr3 and penalized
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Penalized logistic regression object
#> Model failed to converge

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913