Random Ferns Classification Learner

Ensemble machine learning algorithm based on Random Ferns, which are a simplified, faster alternative to Random Forests. Calls rFerns::rFerns() from rFerns.

Initial parameter values

importance:

Actual default: FALSE

Initial value: "simple"

Reason for change: The default value of FALSE will resolve to "none", which turns importance calculation off. To enable importance calculation by default, importance is set to "simple".

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.rFerns")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, rFerns

Parameters

Id	Type	Default	Levels	Range

consistentSeed	untyped	NULL		-

depth	integer	5		\([1, 16]\)

ferns	integer	1000		\((-\infty, \infty)\)

importance	untyped	FALSE		-

saveForest	logical	TRUE	TRUE, FALSE	-

threads	integer	0		\((-\infty, \infty)\)

References

Kursa MB (2014). “rFerns: An Implementation of the Random Ferns Method for General-Purpose Machine Learning.” Journal of Statistical Software, 61(10), 1–13. https://www.jstatsoft.org/v61/i10/.

Ozuysal, Mustafa, Calonder, Michael, Lepetit, Vincent, Fua, Pascal (2010). “Fast Keypoint Recognition Using Random Ferns.” IEEE Transactions on Pattern Analysis and Machine Intelligence, 32(3), 448-461. doi:10.1109/TPAMI.2009.23 .

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifRferns

Methods

Public methods

LearnerClassifRferns$new()

LearnerClassifRferns$importance()

LearnerClassifRferns$oob_error()

LearnerClassifRferns$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifRferns$new()

Method importance()

The importance scores are extracted from the model slot importance.

Usage

LearnerClassifRferns$importance()

Returns

Named numeric().

Method oob_error()

OOB error is extracted from the model slot oobErr.

Usage

LearnerClassifRferns$oob_error()

Returns

numeric(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifRferns$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.rFerns")
print(learner)
#> 
#> ── <LearnerClassifRferns> (classif.rFerns): Random Ferns Classifier ────────────
#> • Model: -
#> • Parameters: importance=simple
#> • Packages: mlr3, mlr3extralearners, and rFerns
#> • Predict Types: [response]
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, multiclass, oob_error, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> 
#>  Forest of 1000 ferns of a depth 5.
#> 
#>  OOB error 21.58%; OOB confusion matrix:
#>          True
#> Predicted  M  R
#>         M 58 14
#>         R 16 51
print(learner$importance())
#>           V11           V12            V9           V10           V36 
#>  0.1001718085  0.0980736487  0.0809824369  0.0746811239  0.0736240957 
#>           V48           V49           V47           V37           V35 
#>  0.0730828475  0.0593247383  0.0495131056  0.0445821114  0.0397165040 
#>           V43           V52           V29           V28           V13 
#>  0.0393038835  0.0391203030  0.0386206549  0.0373358946  0.0372333706 
#>           V46           V51           V21           V20           V39 
#>  0.0340727764  0.0335882093  0.0308945087  0.0301671173  0.0295080918 
#>           V45           V23           V16           V27           V15 
#>  0.0292639497  0.0281022718  0.0261384539  0.0259396024  0.0252207685 
#>           V40           V32           V14           V59            V6 
#>  0.0238717488  0.0233978736  0.0230151105  0.0227687466  0.0227417134 
#>           V58           V19           V38           V18           V24 
#>  0.0224544483  0.0215096367  0.0214481272  0.0207871922  0.0203694492 
#>           V30           V26           V44            V8            V4 
#>  0.0196306979  0.0193936616  0.0188312149  0.0186984265  0.0180391969 
#>           V53           V42           V41            V1           V54 
#>  0.0176454853  0.0170172567  0.0159606134  0.0153126569  0.0150897171 
#>           V31           V34           V25            V2            V5 
#>  0.0149635962  0.0147696990  0.0146164034  0.0145624637  0.0143777699 
#>            V3           V17           V60           V33           V22 
#>  0.0136006240  0.0133549053  0.0128901632  0.0114788981  0.0110522324 
#>           V56           V50           V55            V7           V57 
#>  0.0100822972  0.0084746298  0.0004417407 -0.0004673700 -0.0034887030 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1449275