Classification Random Forest SRC Learner

Random forest for classification. Calls randomForestSRC::rfsrc() from randomForestSRC.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.rfsrc")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3extralearners, randomForestSRC

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

mtry.ratio	numeric	-		\([0, 1]\)

nodesize	integer	15		\([1, \infty)\)

nodedepth	integer	-		\([1, \infty)\)

splitrule	character	gini	gini, auc, entropy	-

nsplit	integer	10		\([0, \infty)\)

importance	character	FALSE	FALSE, TRUE, none, permute, random, anti	-

block.size	integer	10		\([1, \infty)\)

bootstrap	character	by.root	by.root, by.node, none, by.user	-

samptype	character	swor	swor, swr	-

samp	untyped	-		-

membership	logical	FALSE	TRUE, FALSE	-

sampsize	untyped	-		-

sampsize.ratio	numeric	-		\([0, 1]\)

na.action	character	na.omit	na.omit, na.impute	-

nimpute	integer	1		\([1, \infty)\)

proximity	character	FALSE	FALSE, TRUE, inbag, oob, all	-

distance	character	FALSE	FALSE, TRUE, inbag, oob, all	-

forest.wt	character	FALSE	FALSE, TRUE, inbag, oob, all	-

xvar.wt	untyped	-		-

split.wt	untyped	-		-

forest	logical	TRUE	TRUE, FALSE	-

var.used	character	FALSE	FALSE, all.trees	-

split.depth	character	FALSE	FALSE, all.trees, by.tree	-

seed	integer	-		\((-\infty, -1]\)

do.trace	logical	FALSE	TRUE, FALSE	-

get.tree	untyped	-		-

outcome	character	train	train, test	-

ptn.count	integer	0		\([0, \infty)\)

cores	integer	1		\([1, \infty)\)

save.memory	logical	FALSE	TRUE, FALSE	-

perf.type	character	-	gmean, misclass, brier, none	-

case.depth	logical	FALSE	TRUE, FALSE	-

marginal.xvar	untyped	NULL		-

Custom mlr3 parameters

mtry: This hyperparameter can alternatively be set via the added hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

sampsize: This hyperparameter can alternatively be set via the added hyperparameter sampsize.ratio as sampsize = max(ceiling(sampsize.ratio * n_obs), 1). Note that sampsize and sampsize.ratio are mutually exclusive.

cores: This value is set as the option rf.cores during training and is set to 1 by default.

References

Breiman, Leo (2001). “Random Forests.” Machine Learning, 45(1), 5–32. ISSN 1573-0565, doi:10.1023/A:1010933404324 .

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifRandomForestSRC

Methods

Public methods

LearnerClassifRandomForestSRC$new()

LearnerClassifRandomForestSRC$importance()

LearnerClassifRandomForestSRC$selected_features()

LearnerClassifRandomForestSRC$oob_error()

LearnerClassifRandomForestSRC$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifRandomForestSRC$new()

Method importance()

The importance scores are extracted from the model slot importance, returned for 'all'.

Usage

LearnerClassifRandomForestSRC$importance()

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot var.used.

Note: Due to a known issue in randomForestSRC, enabling var.used = "all.trees" causes prediction to fail. Therefore, this setting should be used exclusively for feature selection purposes and not when prediction is required.

Usage

LearnerClassifRandomForestSRC$selected_features()

Returns

character().

Method oob_error()

OOB error extracted from the model slot err.rate.

Usage

LearnerClassifRandomForestSRC$oob_error()

Returns

numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifRandomForestSRC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.rfsrc", importance = "TRUE")
print(learner)
#> 
#> ── <LearnerClassifRandomForestSRC> (classif.rfsrc): Random Forest ──────────────
#> • Model: -
#> • Parameters: importance=TRUE
#> • Packages: mlr3, mlr3extralearners, and randomForestSRC
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, oob_error, selected_features,
#> twoclass, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")
# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#>                          Sample size: 139
#>            Frequency of class labels: M=71, R=68
#>                      Number of trees: 500
#>            Forest terminal node size: 1
#>        Average no. of terminal nodes: 17.588
#> No. of variables tried at each split: 8
#>               Total no. of variables: 60
#>        Resampling used to grow trees: swor
#>     Resample size used to grow trees: 88
#>                             Analysis: RF-C
#>                               Family: class
#>                       Splitting rule: gini *random*
#>        Number of random split points: 10
#>                     Imbalanced ratio: 1.0441
#>                    (OOB) Brier score: 0.14316632
#>         (OOB) Normalized Brier score: 0.57266528
#>                            (OOB) AUC: 0.91228252
#>                       (OOB) Log-loss: 0.45144285
#>                         (OOB) PR-AUC: 0.91135277
#>                         (OOB) G-mean: 0.78932396
#>    (OOB) Requested performance error: 0.20143885, 0.09859155, 0.30882353
#> 
#> Confusion matrix:
#> 
#>           predicted
#>   observed  M  R class.error
#>          M 64  7      0.0986
#>          R 21 47      0.3088
#> 
#>       (OOB) Misclassification rate: 0.2014388
#> 
#> Random-classifier baselines (uniform):
#>    Brier: 0.25   Normalized Brier: 1   Log-loss: 0.69314718
print(learner$importance())
#>            V9           V11           V12           V36           V37 
#>  0.0559038304  0.0535578457  0.0469171080  0.0303088708  0.0292111143 
#>           V47           V10           V49           V21           V45 
#>  0.0246455297  0.0235643232  0.0233672941  0.0215859781  0.0204067077 
#>           V48           V17           V16           V51            V4 
#>  0.0192930767  0.0191604360  0.0187205072  0.0159077605  0.0156122068 
#>            V8           V18           V19           V13           V20 
#>  0.0155562226  0.0153017118  0.0150534352  0.0147001695  0.0136033607 
#>           V46           V35           V30           V23           V15 
#>  0.0135809081  0.0126631985  0.0120871905  0.0116520408  0.0116273746 
#>           V43           V22            V5           V28           V26 
#>  0.0112351828  0.0110494640  0.0097064419  0.0094915498  0.0094728247 
#>            V6           V32           V33            V7           V58 
#>  0.0094464431  0.0087617407  0.0085389803  0.0071313263  0.0069569736 
#>            V2           V44           V39           V34           V14 
#>  0.0068372076  0.0067863443  0.0066955084  0.0059773857  0.0059752723 
#>            V1           V27           V38           V40           V41 
#>  0.0057337677  0.0057226615  0.0057003417  0.0055487139  0.0054206559 
#>           V54           V42           V52           V29            V3 
#>  0.0052630740  0.0049646035  0.0048018576  0.0039529593  0.0039268414 
#>           V55           V25           V31           V59           V50 
#>  0.0036228689  0.0034655564  0.0033319237  0.0031481668  0.0026138163 
#>           V60           V56           V24           V53           V57 
#>  0.0017506033  0.0011508461  0.0007030591 -0.0011863860 -0.0014288175 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1304348