Classification Support Vector Machine Learner

Support Vector classifier trained with John Platt's sequential minimal optimization algorithm. Calls RWeka::SMO() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

E_poly:

original id: E

L_poly:

original id: L

C_poly:

original id: C

C_logistic:

original id: C

R_logistic:

original id: R

M_logistic:

original id: M

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.smo")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

no_checks	logical	FALSE	TRUE, FALSE	-

C	numeric	1		\((-\infty, \infty)\)

N	character	0	0, 1, 2	-

L	numeric	0.001		\((-\infty, \infty)\)

P	numeric	1e-12		\((-\infty, \infty)\)

M	logical	FALSE	TRUE, FALSE	-

V	integer	-1		\((-\infty, \infty)\)

W	integer	1		\((-\infty, \infty)\)

K	character	PolyKernel	NormalizedPolyKernel, PolyKernel, Puk, RBFKernel, StringKernel	-

calibrator	untyped	"weka.classifiers.functions.Logistic"		-

E_poly	numeric	1		\((-\infty, \infty)\)

L_poly	logical	FALSE	TRUE, FALSE	-

C_poly	integer	25007		\((-\infty, \infty)\)

C_logistic	logical	FALSE	TRUE, FALSE	-

R_logistic	numeric	-		\((-\infty, \infty)\)

M_logistic	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

Platt J (1998). “Fast Training of Support Vector Machines using Sequential Minimal Optimization.” In Schoelkopf B, Burges C, Smola A (eds.), Advances in Kernel Methods - Support Vector Learning. MIT Press. http://research.microsoft.com/jplatt/smo.html.

Keerthi S, Shevade S, Bhattacharyya C, Murthy K (2001). “Improvements to Platt's SMO Algorithm for SVM Classifier Design.” Neural Computation, 13(3), 637-649.

Hastie T, Tibshirani R (1998). “Classification by Pairwise Coupling.” In Jordan MI, Kearns MJ, Solla SA (eds.), Advances in Neural Information Processing Systems, volume 10.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSMO

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifSMO$new()

LearnerClassifSMO$marshal()

LearnerClassifSMO$unmarshal()

LearnerClassifSMO$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSMO$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifSMO$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifSMO$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSMO$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.smo")
print(learner)
#> 
#> ── <LearnerClassifSMO> (classif.smo): Support Vector Machine ───────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> SMO
#> 
#> Kernel used:
#>   Linear Kernel: K(x,y) = <x,y>
#> 
#> Classifier for classes: M, R
#> 
#> BinarySMO
#> 
#> Machine linear: showing attribute weights, not support vectors.
#> 
#>         -0.696  * (normalized) V1
#>  +      -0.9883 * (normalized) V10
#>  +      -1.3031 * (normalized) V11
#>  +      -1.0999 * (normalized) V12
#>  +       0.1755 * (normalized) V13
#>  +       0.009  * (normalized) V14
#>  +       0.3846 * (normalized) V15
#>  +       1.0911 * (normalized) V16
#>  +       0.5063 * (normalized) V17
#>  +      -0.1424 * (normalized) V18
#>  +      -0.4555 * (normalized) V19
#>  +      -0.0911 * (normalized) V2
#>  +      -0.6286 * (normalized) V20
#>  +      -0.2794 * (normalized) V21
#>  +       0.3616 * (normalized) V22
#>  +      -0.8761 * (normalized) V23
#>  +      -0.6734 * (normalized) V24
#>  +       0.5789 * (normalized) V25
#>  +       0.8779 * (normalized) V26
#>  +      -0.2996 * (normalized) V27
#>  +      -1.0642 * (normalized) V28
#>  +      -0.3587 * (normalized) V29
#>  +       0.12   * (normalized) V3
#>  +      -0.3913 * (normalized) V30
#>  +       0.9554 * (normalized) V31
#>  +      -0.3442 * (normalized) V32
#>  +       0.3383 * (normalized) V33
#>  +       0.0704 * (normalized) V34
#>  +      -0.0728 * (normalized) V35
#>  +       1.2342 * (normalized) V36
#>  +       1.2324 * (normalized) V37
#>  +      -0.2412 * (normalized) V38
#>  +      -0.7279 * (normalized) V39
#>  +      -1.1693 * (normalized) V4
#>  +       0.7636 * (normalized) V40
#>  +      -0.2166 * (normalized) V41
#>  +       0.0451 * (normalized) V42
#>  +      -0.2029 * (normalized) V43
#>  +      -1.239  * (normalized) V44
#>  +      -1.0818 * (normalized) V45
#>  +       0.0457 * (normalized) V46
#>  +      -0.1348 * (normalized) V47
#>  +      -0.1188 * (normalized) V48
#>  +      -0.8643 * (normalized) V49
#>  +      -0.4906 * (normalized) V5
#>  +       0.9083 * (normalized) V50
#>  +      -0.7139 * (normalized) V51
#>  +      -0.9482 * (normalized) V52
#>  +       0.2368 * (normalized) V53
#>  +       0.1457 * (normalized) V54
#>  +       0.3974 * (normalized) V55
#>  +      -0.1762 * (normalized) V56
#>  +      -0.2043 * (normalized) V57
#>  +      -0.5164 * (normalized) V58
#>  +      -0.5976 * (normalized) V59
#>  +      -0.1549 * (normalized) V6
#>  +       0.0592 * (normalized) V60
#>  +       0.4563 * (normalized) V7
#>  +       0.1902 * (normalized) V8
#>  +      -1.0662 * (normalized) V9
#>  +       2.5015
#> 
#> Number of kernel evaluations: 7279 (87.61% cached)
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203