Classification Logistic Regression Learner

Logistic regression with a quadratic penalization on the coefficient. Calls stepPlr::plr() from stepPlr.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.stepPlr")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, stepPlr

Parameters

Id	Type	Default	Levels	Range

cp	character	aic	aic, bic	-

lambda	numeric	1e-04		\([0, \infty)\)

offset.coefficients	untyped	-		-

offset.subset	untyped	-		-

References

Park, Young M, Hastie, Trevor (2007). “Penalized logistic regression for detecting gene interactions.” Biostatistics, 9(1), 30-50. ISSN 1465-4644, doi:10.1093/biostatistics/kxm010 , https://doi.org/10.1093/biostatistics/kxm010.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifStepPlr

Methods

Public methods

LearnerClassifStepPlr$new()

LearnerClassifStepPlr$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifStepPlr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifStepPlr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.stepPlr")
print(learner)
#> 
#> ── <LearnerClassifStepPlr> (classif.stepPlr): Logistic Regression with a L2 Pena
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and stepPlr
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: twoclass and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> 
#> Call:
#> stepPlr::plr(x = data, y = y)
#> 
#> Coefficients:
#> Intercept        V1       V10       V11       V12       V13       V14       V15 
#>  18.49056 -12.50805   3.60078   2.91736 -30.22697   7.70888 -24.25594  27.40489 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#>  -5.25463  50.74704 -64.07158  33.98460   6.40148 -10.21599 -12.43234  -1.01237 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#>  -1.06943 -45.75006  40.39283  23.14362 -43.12586  34.19027 -24.25016  28.02319 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#>   9.05237  24.16426 -35.61369  -3.57402  31.39854 -19.38779  20.89708   6.13779 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> -11.68441  -2.19180 -49.59037  17.20606  -3.54281 -11.19558  -8.31627  10.27746 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> -38.79316 -13.09082   5.14380 -41.75291 -13.96684  13.26739   6.98545 -38.33580 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> -31.26369 -18.19722 -11.09750   2.55410  -0.13760  -5.20923  -1.67120 -12.46519 
#>        V6       V60        V7        V8        V9 
#>   8.98248 -10.86661  14.44663  20.08080 -27.53941 
#> 
#>     Null deviance: 192.11 on 138 degrees of freedom
#> Residual deviance: 5.26 on 98.15 degrees of freedom
#>             Score: deviance + 4.9 * df = 206.82 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623