Random Forest Competing Risks Learner

Random survival forests for competing risks. Calls randomForestSRC::rfsrc() from randomForestSRC.

Dictionary

This Learner can be instantiated via lrn():

lrn("cmprsk.rfsrc")

Meta Information

Task type: “cmprsk”

Predict Types: “cif”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3cmprsk, mlr3extralearners, randomForestSRC

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

mtry.ratio	numeric	-		\([0, 1]\)

nodesize	integer	15		\([1, \infty)\)

nodedepth	integer	-		\([1, \infty)\)

splitrule	character	logrankCR	logrankCR, logrank	-

nsplit	integer	10		\([0, \infty)\)

importance	character	FALSE	FALSE, TRUE, none, anti, permute, random	-

block.size	integer	10		\([1, \infty)\)

bootstrap	character	by.root	by.root, by.node, none, by.user	-

samptype	character	swor	swor, swr	-

samp	untyped	-		-

membership	logical	FALSE	TRUE, FALSE	-

sampsize	untyped	-		-

sampsize.ratio	numeric	-		\([0, 1]\)

na.action	character	na.omit	na.omit, na.impute	-

nimpute	integer	1		\([1, \infty)\)

ntime	integer	150		\([0, \infty)\)

cause	untyped	-		-

proximity	character	FALSE	FALSE, TRUE, inbag, oob, all	-

distance	character	FALSE	FALSE, TRUE, inbag, oob, all	-

forest.wt	character	FALSE	FALSE, TRUE, inbag, oob, all	-

xvar.wt	untyped	-		-

split.wt	untyped	-		-

forest	logical	TRUE	TRUE, FALSE	-

var.used	character	FALSE	FALSE, all.trees	-

split.depth	character	FALSE	FALSE, all.trees, by.tree	-

seed	integer	-		\((-\infty, -1]\)

do.trace	logical	FALSE	TRUE, FALSE	-

get.tree	untyped	-		-

outcome	character	train	train, test	-

ptn.count	integer	0		\([0, \infty)\)

cores	integer	1		\([1, \infty)\)

save.memory	logical	FALSE	TRUE, FALSE	-

use.uno	logical	TRUE	TRUE, FALSE	-

perf.type	character	-	none	-

case.depth	logical	FALSE	TRUE, FALSE	-

marginal.xvar	untyped	NULL		-

Initial parameter values

ntime: Number of time points to coerce the observed event times for use in the estimated cumulative incidence functions during prediction. We changed the default value of 150 to 0, meaning we now use all the unique event times from the train set across all competing causes.

Custom mlr3 parameters

mtry: This hyperparameter can alternatively be set via the added hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

sampsize: This hyperparameter can alternatively be set via the added hyperparameter sampsize.ratio as sampsize = max(ceiling(sampsize.ratio * n_obs), 1). Note that sampsize and sampsize.ratio are mutually exclusive.

cores: This value is set as the option rf.cores during training and is set to 1 by default.

References

Ishwaran, H., Gerds, A. T, Kogalur, B. U, Moore, D. R, Gange, J. S, Lau, M. B (2014). “Random survival forests for competing risks.” Biostatistics, 15(4), 757–773. doi:10.1093/BIOSTATISTICS/KXU010 , https://doi.org/10.1093/BIOSTATISTICS/KXU010.

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3cmprsk::LearnerCompRisks -> LearnerCompRisksRandomForestSRC

Methods

Public methods

LearnerCompRisksRandomForestSRC$new()

LearnerCompRisksRandomForestSRC$importance()

LearnerCompRisksRandomForestSRC$selected_features()

LearnerCompRisksRandomForestSRC$oob_error()

LearnerCompRisksRandomForestSRC$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerCompRisksRandomForestSRC$new()

Method importance()

The importance scores are extracted from the model slot importance and are cause-specific.

Usage

LearnerCompRisksRandomForestSRC$importance(cause = 1)

Arguments

cause

Integer value indicating the event of interest

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot var.used.

Note: Due to a known issue in randomForestSRC, enabling var.used = "all.trees" causes prediction to fail. Therefore, this setting should be used exclusively for feature selection purposes and not when prediction is required.

Usage

LearnerCompRisksRandomForestSRC$selected_features()

Returns

character().

Method oob_error()

Extracts the out-of-bag (OOB) cumulative incidence function (CIF) error from the model's err.rate slot.

If cause = "mean" (default), the function returns a weighted average of the cause-specific OOB errors, where the weights correspond to the observed proportion of events for each cause in the training data.

Usage

LearnerCompRisksRandomForestSRC$oob_error(cause = "mean")

Arguments

cause

Integer (event type) or "mean" (default). Use a specific event type to retrieve its OOB error, or "mean" to compute the weighted average across causes.

Returns

numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerCompRisksRandomForestSRC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("cmprsk.rfsrc", importance = "TRUE")
print(learner)
#> 
#> ── <LearnerCompRisksRandomForestSRC> (cmprsk.rfsrc): Competing Risk Survival For
#> • Model: -
#> • Parameters: importance=TRUE, ntime=0, cores=1
#> • Packages: mlr3, mlr3cmprsk, mlr3extralearners, and randomForestSRC
#> • Predict Types: [cif]
#> • Feature Types: logical, integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, oob_error, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("pbc")

# Stratification based on event
task$set_col_roles(cols = "status", add_to = "stratum")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#>                          Sample size: 184
#>                     Number of events: 1=12, 2=74
#>                      Number of trees: 500
#>            Forest terminal node size: 15
#>        Average no. of terminal nodes: 9.416
#> No. of variables tried at each split: 5
#>               Total no. of variables: 17
#>        Resampling used to grow trees: swor
#>     Resample size used to grow trees: 116
#>                             Analysis: RSF
#>                               Family: surv-CR
#>                       Splitting rule: logrankCR *random*
#>        Number of random split points: 10
#>    (OOB) Requested performance error: 0.24007791, 0.21101749
#> 
print(learner$importance(cause = 1)) # VIMP for cause = 1
#>          bili           age          chol       protime           ast 
#>  0.2414616727  0.1064792342  0.0608795475  0.0474502977  0.0402365578 
#>         edema      platelet       ascites        copper          trig 
#>  0.0392797958  0.0176260754  0.0143175452  0.0132738782  0.0108224042 
#>      alk.phos           trt       albumin        hepato       spiders 
#>  0.0081005352  0.0029683278  0.0024407952  0.0015136616  0.0011652809 
#>           sex         stage 
#>  0.0004251064 -0.0031654477 
print(learner$importance(cause = 2)) # VIMP for cause = 2
#>          bili         edema        copper           age       albumin 
#>  2.114435e-01  1.075945e-01  9.557630e-02  9.217562e-02  4.394105e-02 
#>       ascites       protime          chol           ast      alk.phos 
#>  4.211369e-02  3.152447e-02  2.355508e-02  2.284940e-02  1.953623e-02 
#>      platelet          trig         stage           sex           trt 
#>  1.179110e-02  1.051888e-02  3.864585e-03 -3.380756e-06 -2.626372e-04 
#>       spiders        hepato 
#> -3.416987e-04 -7.686743e-04 
print(learner$oob_error()) # weighted-mean across causes
#> [1] 0.2150724

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> cmprsk.auc 
#>  0.8895492