Regression Regularized Random Forest Learner

Regularized random forest for regression. Calls RRF::RRF() from RRF.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.RRF")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3extralearners, RRF

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

nodesize	integer	-		\([1, \infty)\)

replace	logical	TRUE	TRUE, FALSE	-

flagReg	integer	1		\([0, \infty)\)

coefReg	numeric	0.8		\((-\infty, \infty)\)

feaIni	untyped	-		-

corr.bias	logical	FALSE	TRUE, FALSE	-

maxnodes	integer	-		\([1, \infty)\)

importance	logical	FALSE	TRUE, FALSE	-

localImp	logical	FALSE	TRUE, FALSE	-

nPerm	integer	1		\([1, \infty)\)

proximity	logical	FALSE	TRUE, FALSE	-

oob.prox	logical	FALSE	TRUE, FALSE	-

do.trace	logical	FALSE	TRUE, FALSE	-

keep.inbag	logical	FALSE	TRUE, FALSE	-

keep.forest	logical	TRUE	TRUE, FALSE	-

strata	untyped	-		-

sampsize	untyped	-		-

predict.all	logical	FALSE	TRUE, FALSE	-

nodes	logical	FALSE	TRUE, FALSE	-

References

Deng, Houtao, Runger, George (2012). “Feature selection via regularized trees.” In 2012 International Joint Conference on Neural Networks (IJCNN), 1–8. IEEE. doi:10.1109/IJCNN.2012.6252640 .

Deng, Houtao, Runger, George (2013). “Gene selection with guided regularized random forest.” Pattern Recognition, 46(12), 3483–3489. doi:10.1016/j.patcog.2013.05.021 .

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrRRF

Methods

Public methods

LearnerRegrRRF$new()

LearnerRegrRRF$importance()

LearnerRegrRRF$oob_error()

LearnerRegrRRF$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrRRF$new()

Method importance()

The importance scores are extracted from the slot importance.

Usage

LearnerRegrRRF$importance()

Returns

Named numeric().

Method oob_error()

OOB errors are extracted from the model slot mse.

Usage

LearnerRegrRRF$oob_error()

Returns

numeric(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrRRF$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.RRF")
print(learner)
#> 
#> ── <LearnerRegrRRF> (regr.RRF): Regularized Random Forest ──────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and RRF
#> • Predict Types: [response]
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance and oob_error
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> 
#> Call:
#>  RRF(formula = task$formula(), data = task$data()) 
#>                Type of random forest: regression
#>                      Number of trees: 500
#> No. of variables tried at each split: 3
#> 
#>           Mean of squared residuals: 8.117204
#>                     % Var explained: 78.33
print(learner$importance())
#>       disp         hp        cyl         wt       drat       qsec       carb 
#> 165.308132 159.440417 104.388391  99.111113  61.961100  28.664940  23.707750 
#>         vs       gear         am 
#>  19.849858   8.789739   8.559714 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.963081