Regression Abess Learner

Adaptive best-subset selection for regression. Calls abess::abess() from abess.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.abess")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, abess

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, mgaussian, poisson, gamma	-

tune.path	character	sequence	sequence, gsection	-

tune.type	character	gic	gic, aic, bic, ebic, cv	-

normalize	integer	NULL		\((-\infty, \infty)\)

support.size	untyped	NULL		-

c.max	integer	2		\([1, \infty)\)

gs.range	untyped	NULL		-

lambda	numeric	0		\([0, \infty)\)

always.include	untyped	NULL		-

group.index	untyped	NULL		-

init.active.set	untyped	NULL		-

splicing.type	integer	2		\([1, 2]\)

max.splicing.iter	integer	20		\([1, \infty)\)

screening.num	integer	NULL		\([0, \infty)\)

important.search	integer	NULL		\([0, \infty)\)

warm.start	logical	TRUE	TRUE, FALSE	-

nfolds	integer	5		\((-\infty, \infty)\)

foldid	untyped	NULL		-

cov.update	logical	FALSE	TRUE, FALSE	-

newton	character	exact	exact, approx	-

newton.thresh	numeric	1e-06		\([0, \infty)\)

max.newton.iter	integer	NULL		\([1, \infty)\)

early.stop	logical	FALSE	TRUE, FALSE	-

ic.scale	numeric	1		\([0, \infty)\)

num.threads	integer	0		\([0, \infty)\)

seed	integer	1		\((-\infty, \infty)\)

Initial parameter values

num.threads: This parameter is initialized to 1 (default is 0) to avoid conflicts with the mlr3 parallelization.

family: Depends on the task type, if the parameter family is NULL, it is set to "binomial" for binary classification tasks and to "multinomial" for multiclass classification problems.

See also

Author

abess-team

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrAbess

Methods

Public methods

LearnerRegrAbess$new()

LearnerRegrAbess$selected_features()

LearnerRegrAbess$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrAbess$new()

Method selected_features()

Extract the name of selected features from the model by abess::extract().

Usage

LearnerRegrAbess$selected_features()

Returns

The names of selected features

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrAbess$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("regr.abess")
print(learner)
#> <LearnerRegrAbess:regr.abess>: Fast Best Subset Selection for Regression
#> * Model: -
#> * Parameters: num.threads=1
#> * Packages: mlr3, abess
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("mtcars")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Call:
#> abess.default(x = task$data(cols = task$feature_names), y = as.matrix(task$data(cols = task$target_names)), 
#>     num.threads = 1L)
#> 
#>   support.size       dev      GIC
#> 1            0 16.998050 59.49507
#> 2            1  3.694682 30.00835
#> 3            2  2.679939 25.82881
#> 4            3  2.409036 26.15447
#> 5            4  2.129968 26.13252
#> 6            5  2.097249 28.37100
#> 7            6  2.088764 30.84944
#> 8            7  2.067916 33.20235
#> 9            8  2.066768 35.75426

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.687728