Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 18.03516                                                 fv = 18.96788                                                 fv = 20.05321                                                 fv = 522.1644                                                 fv = 4.922656                                                 fv = 9614246409                                                 fv = 6287901931                                                 fv = 24440.68                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2800.69                                                 fv = 1.340781e+154                                                 fv = 24440.68                                                 fv = 24440.68                                                 fv = 1.340781e+154                                                 fv = 173.8246                                                 fv = 88448.91                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.922656                                                 fv = 1.340781e+154                                                 fv = 979.3987                                                 fv = 12.18887                                                 fv = 6195797921                                                 fv = 1.340781e+154                                                 fv = 6460118                                                 fv = 64.3481                                                 fv = 123.6031                                                 fv = 1.340781e+154                                                 fv = 4.922656                                                 fv = 4.922656                                                 fv = 558.7649                                                 fv = 4.922656                                                 fv = 32654789                                                 fv = 16.64753                                                 fv = 3919922                                                 fv = 1.340781e+154                                                 fv = 3784592                                                 fv = 3647.527                                                 fv = 17177.99                                                 fv = 11846.36                                                 fv = 1.340781e+154                                                 fv = 26.8331                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 726.6751                                                 fv = 1457.95                                                 fv = 381.0062                                                 fv = 1.340781e+154                                                 fv = 4.922656                                                 fv = 697.7279                                                 fv = 4.922656                                                 fv = 4.922656                                                 fv = 4.922656                                                 fv = 103.9362                                                 fv = 41.92695                                                 fv = 302.683                                                 fv = 36099.02                                                 fv = 11221.66                                                 fv = 40184.85                                                 fv = 42154991                                                 fv = 136.5205 run # 0: f=4.922655988
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.262896                                                 fv = 33.96225                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.878166                                                 fv = 712.3926                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33211478                                                 fv = 33.96225                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5742.817                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37908.66                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.940634                                                 fv = 5.309797                                                 fv = 6.910237                                                 fv = 3.878166                                                 fv = 6091.097                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.475824                                                 fv = 3.878166                                                 fv = 4.155042                                                 fv = 13.8814                                                 fv = 7.469723                                                 fv = 1.340781e+154                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 33.89203                                                 fv = 1.340781e+154                                                 fv = 10.00731                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.458375                                                 fv = 17.68303                                                 fv = 6.1622                                                 fv = 3.878166                                                 fv = 1.340781e+154                                                 fv = 7.1215                                                 fv = 1.340781e+154                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 3.878166                                                 fv = 7.472434                                                 fv = 88.14585                                                 fv = 4.592318                                                 fv = 6.512451                                                 fv = 6.154539                                                 fv = 6.240259                                                 fv = 5.535618                                                 fv = 7.691914                                                 fv = 26.8331                                                 fv = 1.340781e+154                                                 fv = 6.712161                                                 fv = 7.644798                                                 fv = 3.878166                                                 fv = 4.324534                                                 fv = 4.324534                                                 fv = 1.340781e+154                                                 fv = 6.975635                                                 fv = 6.779045                                                 fv = 1.340781e+154                                                 fv = 6.701452                                                 fv = 7.308612                                                 fv = 3.878166                                                 fv = 6.340207                                                 fv = 3.878166                                                 fv = 18.88959                                                 fv = 3.878166                                                 fv = 6.917354                                                 fv = 33.96225 run # 1: f=3.878166009
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.40829                                                 fv = 33.96225                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42654372824                                                 fv = 33.96225                                                 fv = 33.96225                                                 fv = 25.65312                                                 fv = 25.65312                                                 fv = 5479.421                                                 fv = 25.65312                                                 fv = 308998263                                                 fv = 1.340781e+154                                                 fv = 10.18473                                                 fv = 10.92086                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.18473                                                 fv = 10.18473                                                 fv = 10.18473                                                 fv = 10.18473                                                 fv = 1.340781e+154                                                 fv = 10.18473                                                 fv = 10.92086                                                 fv = 1.340781e+154                                                 fv = 12.15095                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 98105086915                                                 fv = 81649677                                                 fv = 4384.002                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.02526                                                 fv = 1.340781e+154                                                 fv = 29.16835                                                 fv = 25.88896                                                 fv = 10.18473                                                 fv = 20.04422                                                 fv = 10.18473                                                 fv = 10.18473                                                 fv = 167.85                                                 fv = 50.20396                                                 fv = 26.22376                                                 fv = 25.65312                                                 fv = 81017.32                                                 fv = 81017.32                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 60.61185                                                 fv = 7.381832                                                 fv = 14291.98                                                 fv = 1.340781e+154                                                 fv = 10.18473                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.381832                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.255923                                                 fv = 6.255923                                                 fv = 1.340781e+154                                                 fv = 10.18473                                                 fv = 3.239923e+12                                                 fv = 7.776843                                                 fv = 12635.12                                                 fv = 1.340781e+154                                                 fv = 8.236386                                                 fv = 6.530033                                                 fv = 6.255923                                                 fv = 7.776843                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.776843                                                 fv = 7.776843                                                 fv = 6.255923                                                 fv = 1.340781e+154                                                 fv = 8.629826                                                 fv = 9291.578                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 76418105                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.239901                                                 fv = 6.255923                                                 fv = 12.26265                                                 fv = 6.255923                                                 fv = 8.236386                                                 fv = 8.236386                                                 fv = 25.65312                                                 fv = 1.340781e+154                                                 fv = 8.236386                                                 fv = 6.255923                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.65312                                                 fv = 5.598814                                                 fv = 302149152                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.598814                                                 fv = 1.340781e+154                                                 fv = 115.9766                                                 fv = 1.340781e+154                                                 fv = 5.598814                                                 fv = 6.311047                                                 fv = 7.344917                                                 fv = 6.030262                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 8.735749                                                 fv = 7.208157                                                 fv = 9.346767                                                 fv = 7.472826                                                 fv = 1.340781e+154                                                 fv = 10.65205                                                 fv = 1.340781e+154                                                 fv = 754.7611                                                 fv = 34.90007                                                 fv = 22.35813                                                 fv = 5.598814                                                 fv = 74.36906                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.214283                                                 fv = 6.146279                                                 fv = 6.396591                                                 fv = 10.77202                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 5.598814                                                 fv = 6.688204                                                 fv = 6.030262                                                 fv = 6.640193                                                 fv = 1.340781e+154                                                 fv = 1895.765                                                 fv = 260.9501                                                 fv = 33.96225                                                 fv = 1.340781e+154                                                 fv = 84.04557                                                 fv = 34.90007                                                 fv = 23.15658                                                 fv = 1.340781e+154                                                 fv = 9.346767                                                 fv = 8.627624 run # 2: f=5.598814492
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1038.108                                                 fv = 33.96225                                                 fv = 33.96225                                                 fv = 177926358                                                 fv = 7.965237                                                 fv = 7.760638                                                 fv = 7.760638                                                 fv = 1.340781e+154                                                 fv = 12.15771                                                 fv = 1.340781e+154                                                 fv = 1474405088                                                 fv = 1474405088                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1563.284                                                 fv = 233.62                                                 fv = 6969.814                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.54113                                                 fv = 1.340781e+154                                                 fv = 6.066395                                                 fv = 6.291241                                                 fv = 7.760638                                                 fv = 8.246626                                                 fv = 1.340781e+154                                                 fv = 6.066395                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.48923 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 5.678834                                                 fv = 1600826                                                 fv = 1.340781e+154                                                 fv = 8.515179                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.513608                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 75.80856                                                 fv = 1.340781e+154                                                 fv = 8.513608                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 90.17892                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 5.678834                                                 fv = 1.340781e+154                                                 fv = 155.0398                                                 fv = 8.513608                                                 fv = 1.340781e+154                                                 fv = 8.246626 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 5.678834                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.653295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.3223                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.170679                                                 fv = 7.787705                                                 fv = 5.653295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.44663                                                 fv = 1.340781e+154                                                 fv = 256.3362                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 167938.3                                                 fv = 1.340781e+154                                                 fv = 6.066395                                                 fv = 106630.3                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11041508                                                 fv = 5.653295                                                 fv = 5.653295                                                 fv = 5.653295                                                 fv = 7.760638                                                 fv = 5.653295                                                 fv = 8.170679                                                 fv = 1.340781e+154                                                 fv = 7.760638                                                 fv = 122.7743                                                 fv = 12.15771                                                 fv = 7.760638                                                 fv = 7.760638                                                 fv = 347.0554                                                 fv = 5.653295                                                 fv = 1.340781e+154                                                 fv = 8.516429                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.066395                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.653295                                                 fv = 5.791401                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.290997                                                 fv = 9.129117                                                 fv = 5.653295                                                 fv = 5.653295                                                 fv = 162.8604                                                 fv = 11.5586                                                 fv = 318.546                                                 fv = 1.340781e+154                                                 fv = 46.15412                                                 fv = 1.340781e+154                                                 fv = 11.5586                                                 fv = 6.446943                                                 fv = 93.90682                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.186903                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.97801                                                 fv = 10.28308                                                 fv = 9.619356                                                 fv = 137.761                                                 fv = 5.653295                                                 fv = 5.653295                                                 fv = 5.653295                                                 fv = 137.761                                                 fv = 103.7541                                                 fv = 8.808125 run # 3: f=5.653295306
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 586
#> best    : 3.878166009
#> worst   : 1.340780793e+154
#> solution: x = ( 1 1 0 0 1 0 0 0 0 0 1 1 2 1 1 1 4 1 1 1 ) f(x) = 3.878166009
#> 
#>                                                 fv = 3.878166                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (5): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 25.02279