Regression Conditional Inference Tree Learner

Regression Partition Tree where a significance test is used to determine the univariate splits. Calls partykit::ctree() from partykit.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.ctree")
lrn("regr.ctree")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, partykit, sandwich, coin

Parameters

Id	Type	Default	Levels	Range

teststat	character	quadratic	quadratic, maximum	-

splitstat	character	quadratic	quadratic, maximum	-

splittest	logical	FALSE	TRUE, FALSE	-

testtype	character	Bonferroni	Bonferroni, MonteCarlo, Univariate, Teststatistic	-

nmax	untyped	-		-

alpha	numeric	0.05		\([0, 1]\)

mincriterion	numeric	0.95		\([0, 1]\)

logmincriterion	numeric	-		\((-\infty, \infty)\)

minsplit	integer	20		\([1, \infty)\)

minbucket	integer	7		\([1, \infty)\)

minprob	numeric	0.01		\([0, \infty)\)

stump	logical	FALSE	TRUE, FALSE	-

lookahead	logical	FALSE	TRUE, FALSE	-

MIA	logical	FALSE	TRUE, FALSE	-

maxvar	integer	-		\([1, \infty)\)

nresample	integer	9999		\([1, \infty)\)

tol	numeric	-		\([0, \infty)\)

maxsurrogate	integer	0		\([0, \infty)\)

numsurrogate	logical	FALSE	TRUE, FALSE	-

mtry	integer	Inf		\([0, \infty)\)

maxdepth	integer	Inf		\([0, \infty)\)

multiway	logical	FALSE	TRUE, FALSE	-

splittry	integer	2		\([0, \infty)\)

intersplit	logical	FALSE	TRUE, FALSE	-

majority	logical	FALSE	TRUE, FALSE	-

caseweights	logical	FALSE	TRUE, FALSE	-

applyfun	untyped	-		-

cores	integer	NULL		\((-\infty, \infty)\)

saveinfo	logical	TRUE	TRUE, FALSE	-

update	logical	FALSE	TRUE, FALSE	-

splitflavour	character	ctree	ctree, exhaustive	-

offset	untyped	-		-

cluster	untyped	-		-

scores	untyped	-		-

doFit	logical	TRUE	TRUE, FALSE	-

maxpts	integer	25000		\((-\infty, \infty)\)

abseps	numeric	0.001		\([0, \infty)\)

releps	numeric	0		\([0, \infty)\)

References

Hothorn T, Zeileis A (2015). “partykit: A Modular Toolkit for Recursive Partytioning in R.” Journal of Machine Learning Research, 16(118), 3905-3909. http://jmlr.org/papers/v16/hothorn15a.html.

Hothorn T, Hornik K, Zeileis A (2006). “Unbiased Recursive Partitioning: A Conditional Inference Framework.” Journal of Computational and Graphical Statistics, 15(3), 651--674. doi:10.1198/106186006x133933 , https://doi.org/10.1198/106186006x133933.

See also

Author

sumny

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCTree

Methods

Public methods

LearnerRegrCTree$new()

LearnerRegrCTree$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCTree$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCTree$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("regr.ctree")
print(learner)
#> <LearnerRegrCTree:regr.ctree>: Conditional Inference Tree
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, partykit, sandwich, coin
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: weights

# available parameters:
learner$param_set$ids()
#>  [1] "teststat"        "splitstat"       "splittest"       "testtype"       
#>  [5] "nmax"            "alpha"           "mincriterion"    "logmincriterion"
#>  [9] "minsplit"        "minbucket"       "minprob"         "stump"          
#> [13] "lookahead"       "MIA"             "maxvar"          "nresample"      
#> [17] "tol"             "maxsurrogate"    "numsurrogate"    "mtry"           
#> [21] "maxdepth"        "multiway"        "splittry"        "intersplit"     
#> [25] "majority"        "caseweights"     "applyfun"        "cores"          
#> [29] "saveinfo"        "update"          "splitflavour"    "offset"         
#> [33] "cluster"         "scores"          "doFit"           "maxpts"         
#> [37] "abseps"          "releps"