Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 37.63758 35.22566 30.77569 26.85108 23.65249 21.07088 19.01237 17.39624
#>   [9] 16.15312 15.22348 14.55627 14.10778 13.84065 13.72296 13.72751 13.77984
#>  [17] 13.47407 13.20022 13.19587 13.24379 13.25997 13.24672 13.21280 13.09120
#>  [25] 12.93604 12.80881 12.70537 14.63457 13.72659 13.61092 14.71507 15.03047
#>  [33] 14.64214 14.62382 14.72499 14.63988 15.67365 15.66309 15.40559 15.37206
#>  [41] 15.30895 15.45337 13.75578 13.70270 14.16648 14.22010 13.87985 17.74992
#>  [49] 17.91209 18.43272 22.17899 22.09815 28.90972 28.91176 29.17824 30.28958
#>  [57] 30.29148 30.29351 31.38327 31.58850 31.75005 31.79748 31.93399 32.69572
#>  [65] 32.86007 32.85980 34.25808 34.25806 34.40799 34.40802 34.47619 34.66327
#>  [73] 34.66315 34.66300 34.32998 34.29695 34.29576 34.29461 34.29349 34.29240
#>  [81] 34.59573 34.59564 34.59574 34.63335 34.63318 34.63289 34.61407 34.61388
#>  [89] 34.61368 34.61347 34.61325 34.61303 34.05802 34.05856 34.05907 34.05954
#>  [97] 34.09950 34.10033 34.10112 34.10188
#> 
#> $cvse
#>   [1] 12.014280 11.957123 10.919401  9.796321  8.803876  7.925236  7.148994
#>   [8]  6.468342  5.880284  5.384738  4.983397  4.678199  4.469492  4.354334
#>  [15]  4.325594  4.354458  4.230666  4.116289  4.106938  4.101309  4.078668
#>  [22]  4.047074  4.024770  3.995021  3.936085  3.884033  3.837977  4.361492
#>  [29]  4.117365  4.083913  4.084909  4.116101  4.074736  4.069596  4.044272
#>  [36]  4.054904  4.138406  4.139391  4.173313  4.151586  4.111122  4.255590
#>  [43]  3.960991  3.969640  4.247506  4.254014  4.280375  6.750823  6.736520
#>  [50]  6.674023  7.663549  7.654025 12.584440 12.585733 12.558811 12.470985
#>  [57] 12.472155 12.473303 12.640384 12.642773 12.627045 12.602054 12.592782
#>  [64] 12.574779 12.558963 12.557888 12.553528 12.552533 12.568409 12.567499
#>  [71] 12.558977 12.585528 12.584726 12.583946 12.581770 12.581808 12.581092
#>  [78] 12.580398 12.579725 12.579074 12.580478 12.579874 12.579227 12.582744
#>  [85] 12.582170 12.581612 12.581234 12.580717 12.580216 12.579730 12.579259
#>  [92] 12.578801 11.754096 11.753788 11.753483 11.753184 11.754011 11.753746
#>  [99] 11.753485 11.753227
#> 
#> $fold
#>  [1]  3 10  1  4  2  5  8  2  9  7  5  4  3  6  1  6 10  8  1  7  9
#> 
#> $lambda
#>   [1] 5.176524678 4.827644240 4.502277175 4.198838762 3.915851083 3.651935826
#>   [7] 3.405807574 3.176267542 2.962197740 2.762555527 2.576368530 2.402729914
#>  [13] 2.240793959 2.089771946 1.948928311 1.817577066 1.695078455 1.580835840
#>  [19] 1.474292796 1.374930395 1.282264689 1.195844340 1.115248434 1.040084422
#>  [25] 0.969986213 0.904612389 0.843644541 0.786785723 0.733758998 0.684306098
#>  [31] 0.638186157 0.595174546 0.555061774 0.517652468 0.482764424 0.450227718
#>  [37] 0.419883876 0.391585108 0.365193583 0.340580757 0.317626754 0.296219774
#>  [43] 0.276255553 0.257636853 0.240272991 0.224079396 0.208977195 0.194892831
#>  [49] 0.181757707 0.169507845 0.158083584 0.147429280 0.137493040 0.128226469
#>  [55] 0.119584434 0.111524843 0.104008442 0.096998621 0.090461239 0.084364454
#>  [61] 0.078678572 0.073375900 0.068430610 0.063818616 0.059517455 0.055506177
#>  [67] 0.051765247 0.048276442 0.045022772 0.041988388 0.039158511 0.036519358
#>  [73] 0.034058076 0.031762675 0.029621977 0.027625555 0.025763685 0.024027299
#>  [79] 0.022407940 0.020897719 0.019489283 0.018175771 0.016950785 0.015808358
#>  [85] 0.014742928 0.013749304 0.012822647 0.011958443 0.011152484 0.010400844
#>  [91] 0.009699862 0.009046124 0.008436445 0.007867857 0.007337590 0.006843061
#>  [97] 0.006381862 0.005951745 0.005550618 0.005176525
#> 
#> $fit
#> $beta
#>               5.1765     4.8276    4.5023    4.1988    3.9159    3.6519
#> (Intercept) 20.08571 21.7419111 23.286486 24.726961 26.070354 27.323206
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5174614 -1.000048 -1.450109 -1.869838 -2.261279
#>                3.4058    3.1763    2.9622    2.7626    2.5764    2.4027
#> (Intercept) 28.491620 29.581287 30.597514 31.545251 32.429113 33.253406
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.626337 -2.966793 -3.284302 -3.580413 -3.856566 -4.114108
#>                2.2408    2.0898    1.9489   1.8176    1.6951    1.5808
#> (Intercept) 34.022145 34.739073 35.407682 36.03123 36.468297 36.468297
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt          -4.354292 -4.578289 -4.787189 -4.98201 -5.118567 -5.118567
#>                1.4743    1.3749    1.2823      1.1958     1.1152     1.0401
#> (Intercept) 36.468297 36.468297 36.468297 35.09439904 33.5860431 32.1792877
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.07312364  0.1534036  0.2282761
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.118567 -5.118567 -5.118567 -5.09468376 -5.0684633 -5.0440091
#>                 0.9700     0.9046     0.8436     0.7868     0.7338     0.6843
#> (Intercept) 30.8673429 29.6438187 28.5027560 27.4385971 26.4461590 25.5206078
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2981023  0.3632226  0.4239539  0.4805922  0.5334132  0.5826743
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.0212030 -4.9999339 -4.9800983 -4.9615995 -4.9443475 -4.9282583
#>                 0.6382     0.5952     0.5551     0.5177     0.4828    0.4502
#> (Intercept) 24.6574358 23.8524386 23.1016955 22.4015500 21.7485920  8.087421
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  3.036840
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.6286153  0.6714601  0.7114172  0.7486814  0.7834341  1.267780
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -4.9132534 -4.8992598 -4.8862092 -4.8740383 -4.8626876 -3.686113
#>                0.4199    0.3916    0.3652    0.3406    0.3176    0.2962
#> (Intercept)  8.083625  8.080908  8.078962  8.077568  8.076570  8.075856
#> am           3.037658  3.038244  3.038664  3.038964  3.039179  3.039333
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.267916  1.268014  1.268084  1.268134  1.268170  1.268196
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.685794 -3.685566 -3.685402 -3.685285 -3.685201 -3.685141
#>                 0.2763     0.2576     0.2403     0.2241     0.2090     0.1949
#> (Intercept) 14.3955163 14.3930658 12.6374018 11.1602927 11.1636409 11.1661353
#> am           3.3942257  3.3950069  3.0936058  2.8359960  2.8353694  2.8349294
#> carb        -0.4945260 -0.4946009 -0.5547217 -0.6056268 -0.6057258 -0.6057958
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.6917538  1.2774618  1.2776290  1.2777257
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9203971  0.9204677  0.8640650  0.8160009  0.8158336  0.8157138
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.3155477 -3.3152076 -3.1309243 -2.9758010 -2.9759301 -2.9760290
#>                 0.1818     0.1695     0.1581     0.1474     0.1375     0.1282
#> (Intercept) 11.1679627 11.1692838 11.1702286 11.1708980 11.1713688 21.0333587
#> am           2.8346225  2.8344096  2.8342626  2.8341616  2.8340924  2.7247114
#> carb        -0.6058450 -0.6058797 -0.6059039 -0.6059210 -0.6059329 -0.5107151
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.2777810  1.2778125  1.2778302  1.2778402  1.2778458  0.9843320
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -0.0156581
#> qsec         0.8156287  0.8155686  0.8155264  0.8154969  0.8154763  0.3787866
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9761027 -2.9761564 -2.9761949 -2.9762222 -2.9762412 -2.6314369
#>                  0.1196      0.1115      0.1040      0.0970      0.0905
#> (Intercept) 21.03654280 21.03891866 21.04066440 16.64139150 16.63799492
#> am           2.72476835  2.72481906  2.72486375  2.74849633  2.74840323
#> carb        -0.51053715 -0.51039409 -0.51027965  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.01469169  0.01468921
#> drat         0.98386257  0.98348614  0.98318580  0.77095615  0.77112913
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01566412 -0.01566869 -0.01567211 -0.02575254 -0.02574681
#> qsec         0.37872900  0.37869171  0.37866948  0.74269101  0.74281705
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -2.63147288 -2.63150924 -2.63154456 -4.08971019 -4.08962550
#>                  0.0844      0.0787      0.0734      0.0684      0.0638
#> (Intercept) 16.63521582 16.63295927 16.63114268 16.62969452 16.62855318
#> am           2.74831150  2.74822293  2.74813876  2.74805982  2.74798661
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01468699  0.01468503  0.01468331  0.01468180  0.01468048
#> drat         0.77128296  0.77141900  0.77153866  0.77164337  0.77173455
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02574207 -0.02573817 -0.02573499 -0.02573241 -0.02573034
#> qsec         0.74291715  0.74299571  0.74305647  0.74310262  0.74313689
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.08954076 -4.08945802 -4.08937874 -4.08930391 -4.08923416
#>                  0.0595      0.0555      0.0518      0.0483      0.0450
#> (Intercept) 16.62766579 16.62698724 16.62647922 16.62610935 16.62585039
#> am           2.74791936  2.74785810  2.74780273  2.74775301  2.74770866
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01467934  0.01467835  0.01467750  0.01467678  0.01467616
#> drat         0.77181356  0.77188169  0.77194016  0.77199011  0.77203256
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02572870 -0.02572741 -0.02572641 -0.02572565 -0.02572509
#> qsec         0.74316155  0.74317852  0.74318941  0.74319553  0.74319800
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.08916983 -4.08911105 -4.08905778 -4.08900984 -4.08896699
#>                  0.0420      0.0392      0.0365      0.0341      0.0318
#> (Intercept) 16.62567959 16.62557801 16.62553002 16.62552278 16.62554584
#> am           2.74766933  2.74763464  2.74760421  2.74757764  2.74755457
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01467564  0.01467521  0.01467484  0.01467453  0.01467428
#> drat         0.77206847  0.77209868  0.77212396  0.77214499  0.77216239
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02572468 -0.02572441 -0.02572423 -0.02572412 -0.02572408
#> qsec         0.74319770  0.74319536  0.74319157  0.74318681  0.74318143
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.08892893 -4.08889532 -4.08886580 -4.08884000 -4.08881757
#>                  0.0296      0.0276      0.0258      0.0240      0.0224
#> (Intercept) 16.62559073 16.62565070 16.62572037 16.61955725  9.69192983
#> am           2.74753463  2.74751748  2.74750280  2.74683769  2.18258088
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01467407  0.01467390  0.01467376  0.01467351  0.01676108
#> drat         0.77217668  0.77218833  0.77219777  0.77216637  0.80184689
#> gear         0.00000000  0.00000000  0.00000000  0.00103668  0.64118136
#> hp          -0.02572409 -0.02572412 -0.02572419 -0.02572604 -0.02580128
#> qsec         0.74317572  0.74316991  0.74316417  0.74327785  1.06172960
#> vs           0.00000000  0.00000000  0.00000000  0.00000000 -0.87405995
#> wt          -4.08879817 -4.08878147 -4.08876716 -4.08844290 -4.40606730
#>                  0.0209      0.0195      0.0182      0.0170      0.0158
#> (Intercept)  9.68481103  9.67796549  9.67138270  9.66505258  9.65896543
#> am           2.18213529  2.18170681  2.18129478  2.18089858  2.18051758
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01676348  0.01676579  0.01676801  0.01677015  0.01677221
#> drat         0.80187768  0.80190727  0.80193573  0.80196310  0.80198941
#> gear         0.64169418  0.64218732  0.64266153  0.64311754  0.64355605
#> hp          -0.02579934 -0.02579748 -0.02579568 -0.02579396 -0.02579230
#> qsec         1.06207184  1.06240094  1.06271741  1.06302173  1.06331437
#> vs          -0.87479124 -0.87549447 -0.87617070 -0.87682097 -0.87744629
#> wt          -4.40646698 -4.40685133 -4.40722094 -4.40757636 -4.40791815
#>                  0.0147      0.0137      0.0128      0.0120      0.0112
#> (Intercept)  9.65311191  9.64748306  9.64207024  9.63686517  9.63185987
#> am           2.18015121  2.17979891  2.17946013  2.17913436  2.17882109
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01677418  0.01677608  0.01677791  0.01677967  0.01678136
#> drat         0.80201471  0.80203903  0.80206242  0.80208491  0.80210654
#> gear         0.64397772  0.64438321  0.64477313  0.64514809  0.64550865
#> hp          -0.02579071 -0.02578918 -0.02578770 -0.02578628 -0.02578492
#> qsec         1.06359579  1.06386640  1.06412662  1.06437686  1.06461750
#> vs          -0.87804760 -0.87862584 -0.87918188 -0.87971658 -0.88023076
#> wt          -4.40824682 -4.40856288 -4.40886681 -4.40915907 -4.40944012
#>                  0.0104      0.0097      0.0090        0.0084      0.0079
#> (Intercept)  9.62704667  9.62241820  9.61796736  9.6145612817 10.00565723
#> am           2.17851985  2.17823017  2.17795161  2.1776837485  2.06993367
#> carb         0.00000000  0.00000000  0.00000000  0.0000000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000 -0.0002323469 -0.09122249
#> disp         0.01678298  0.01678455  0.01678605  0.0167903438  0.01707011
#> drat         0.80212734  0.80214734  0.80216657  0.8021552083  0.80498331
#> gear         0.64585537  0.64618879  0.64650941  0.6468671342  0.67271440
#> hp          -0.02578361 -0.02578235 -0.02578114 -0.0257792901 -0.02528943
#> qsec         1.06484890  1.06507142  1.06528539  1.0654825502  1.06680636
#> vs          -0.88072520 -0.88120067 -0.88165789 -0.8821484165 -1.02020275
#> wt          -4.40971038 -4.40997028 -4.41022019 -4.4104830268 -4.40309956
#>                  0.0073      0.0068      0.0064      0.0060      0.0056
#> (Intercept) 10.00193326  9.99813648  9.99428745  9.99040501  9.98650634
#> am           2.06889303  2.06792838  2.06703477  2.06620737  2.06544148
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -0.09141527 -0.09157799 -0.09171445 -0.09182815 -0.09192225
#> disp         0.01707092  0.01707191  0.01707305  0.01707433  0.01707573
#> drat         0.80513267  0.80526156  0.80537171  0.80546483  0.80554255
#> gear         0.67338597  0.67403404  0.67465845  0.67525918  0.67583638
#> hp          -0.02529088 -0.02529236 -0.02529384 -0.02529529 -0.02529669
#> qsec         1.06697652  1.06714843  1.06732170  1.06749596  1.06767083
#> vs          -1.02122299 -1.02217110 -1.02305325 -1.02387527 -1.02464257
#> wt          -4.40314474 -4.40321774 -4.40331568 -4.40343578 -4.40357536
#>                  0.0052
#> (Intercept)  9.80946858
#> am           2.03308281
#> carb        -0.02555345
#> cyl         -0.09831724
#> disp         0.01683922
#> drat         0.81937195
#> gear         0.71334328
#> hp          -0.02486222
#> qsec         1.06966700
#> vs          -1.07063758
#> wt          -4.37451359
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3  47
#>  [37]   1   1   1   1   1   1  59   1  57  29   1   1   1   1   1   1   1  37
#>  [55]   1   1   1  79   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   2 131   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1  42   1   1   1   1   1  41
#> 
#> $lambda
#>   [1] 5.176524678 4.827644240 4.502277175 4.198838762 3.915851083 3.651935826
#>   [7] 3.405807574 3.176267542 2.962197740 2.762555527 2.576368530 2.402729914
#>  [13] 2.240793959 2.089771946 1.948928311 1.817577066 1.695078455 1.580835840
#>  [19] 1.474292796 1.374930395 1.282264689 1.195844340 1.115248434 1.040084422
#>  [25] 0.969986213 0.904612389 0.843644541 0.786785723 0.733758998 0.684306098
#>  [31] 0.638186157 0.595174546 0.555061774 0.517652468 0.482764424 0.450227718
#>  [37] 0.419883876 0.391585108 0.365193583 0.340580757 0.317626754 0.296219774
#>  [43] 0.276255553 0.257636853 0.240272991 0.224079396 0.208977195 0.194892831
#>  [49] 0.181757707 0.169507845 0.158083584 0.147429280 0.137493040 0.128226469
#>  [55] 0.119584434 0.111524843 0.104008442 0.096998621 0.090461239 0.084364454
#>  [61] 0.078678572 0.073375900 0.068430610 0.063818616 0.059517455 0.055506177
#>  [67] 0.051765247 0.048276442 0.045022772 0.041988388 0.039158511 0.036519358
#>  [73] 0.034058076 0.031762675 0.029621977 0.027625555 0.025763685 0.024027299
#>  [79] 0.022407940 0.020897719 0.019489283 0.018175771 0.016950785 0.015808358
#>  [85] 0.014742928 0.013749304 0.012822647 0.011958443 0.011152484 0.010400844
#>  [91] 0.009699862 0.009046124 0.008436445 0.007867857 0.007337590 0.006843061
#>  [97] 0.006381862 0.005951745 0.005550618 0.005176525
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 35
#> 
#> $loss
#>   [1] 724.20571 616.17961 525.79972 450.52660 388.16784 336.83173 294.88708
#>   [8] 260.92835 233.74535 212.29697 195.68842 183.15138 174.02694 167.75066
#>  [15] 163.83971 161.88167 161.48115 161.48115 161.48115 161.48115 161.48115
#>  [22] 155.49214 149.55722 144.62596 140.55230 137.21002 134.49034 132.29952
#>  [29] 130.55693 129.19320 128.14874 127.37244 126.82044 126.45524 126.24477
#>  [36] 111.20172 111.20172 111.20171 111.20171 111.20171 111.20171 111.20171
#>  [43] 105.92303 105.92303 103.02491 102.25281 102.25281 102.25281 102.25280
#>  [50] 102.25280 102.25280 102.25280 102.25280  99.03461  99.03461  99.03460
#>  [57]  99.03460  95.17713  95.17712  95.17712  95.17712  95.17711  95.17711
#>  [64]  95.17711  95.17711  95.17711  95.17711  95.17711  95.17711  95.17711
#>  [71]  95.17711  95.17711  95.17711  95.17711  95.17711  95.17711  95.17711
#>  [78]  95.17630  94.74078  94.74076  94.74074  94.74073  94.74071  94.74070
#>  [85]  94.74069  94.74067  94.74066  94.74065  94.74064  94.74063  94.74063
#>  [92]  94.74062  94.74047  94.71149  94.71147  94.71145  94.71144  94.71142
#>  [99]  94.71141  94.70535
#> 
#> $linear.predictors
#>      5.1765   4.8276   4.5023   4.1988   3.9159   3.6519   3.4058   3.1763
#> 1  20.08571 20.25421 20.41135 20.55790 20.69457 20.82203 20.94090 21.05176
#> 2  20.08571 20.54140 20.96638 21.36271 21.73233 22.07704 22.39852 22.69833
#> 3  20.08571 20.07827 20.07133 20.06486 20.05882 20.05319 20.04795 20.04305
#> 4  20.08571 19.95149 19.82632 19.70958 19.60071 19.49918 19.40449 19.31618
#> 5  20.08571 19.89457 19.71632 19.55007 19.39503 19.25044 19.11560 18.98984
#> 6  20.08571 19.96184 19.84632 19.73859 19.63811 19.54441 19.45702 19.37552
#> 7  20.08571 19.96184 19.84632 19.73859 19.63811 19.54441 19.45702 19.37552
#> 8  20.08571 19.63584 19.21629 18.82502 18.46011 18.11980 17.80243 17.50644
#> 9  20.08571 19.81178 19.55631 19.31805 19.09586 18.88864 18.69538 18.51515
#> 10 20.08571 19.78591 19.50631 19.24555 19.00237 18.77557 18.56406 18.36681
#> 11 20.08571 19.02524 18.03624 17.11389 16.25370 15.45149 14.70335 14.00563
#> 12 20.08571 18.97608 17.94123 16.97613 16.07607 15.23667 14.45385 13.72378
#> 13 20.08571 20.90621 21.67141 22.38504 23.05057 23.67124 24.25009 24.78992
#> 14 20.08571 20.79237 21.45140 22.06601 22.63920 23.17376 23.67229 24.13722
#> 15 20.08571 20.46637 20.82137 21.15244 21.46120 21.74915 22.01770 22.26814
#> 16 20.08571 19.75227 19.44130 19.15129 18.88083 18.62859 18.39335 18.17397
#> 17 20.08571 20.74062 21.35139 21.92100 22.45222 22.94763 23.40966 23.84054
#> 18 20.08571 20.95899 21.77341 22.53295 23.24129 23.90189 24.51797 25.09253
#> 19 20.08571 20.10156 20.11633 20.13012 20.14297 20.15495 20.16613 20.17655
#> 20 20.08571 20.30854 20.51635 20.71016 20.89090 21.05946 21.21667 21.36327
#> 21 20.08571 19.89457 19.71632 19.55007 19.39503 19.25044 19.11560 18.98984
#>      2.9622   2.7626   2.5764   2.4027   2.2408   2.0898    1.9489    1.8176
#> 1  21.15515 21.25156 21.34149 21.42535 21.50355 21.57649 21.644514 21.707952
#> 2  22.97793 23.23869 23.48188 23.70868 23.92019 24.11744 24.301404 24.472967
#> 3  20.03848 20.03422 20.03025 20.02655 20.02310 20.01987 20.016870 20.014068
#> 4  19.23383 19.15702 19.08539 19.01859 18.95629 18.89819 18.844009 18.793476
#> 5  18.87256 18.76318 18.66117 18.56604 18.47732 18.39458 18.317418 18.245455
#> 6  19.29951 19.22863 19.16253 19.10088 19.04338 18.98976 18.939752 18.893116
#> 7  19.29951 19.22863 19.16253 19.10088 19.04338 18.98976 18.939752 18.893116
#> 8  17.23040 16.97297 16.73289 16.50899 16.30018 16.10544 15.923823 15.754450
#> 9  18.34707 18.19031 18.04412 17.90778 17.78063 17.66206 17.551468 17.448333
#> 10 18.18285 18.01129 17.85129 17.70208 17.56292 17.43314 17.312108 17.199233
#> 11 13.35493 12.74809 12.18214 11.65434 11.16211 10.70306 10.274940  9.875678
#> 12 13.04292 12.40795 11.81577 11.26350 10.74845 10.26812  9.820158  9.402387
#> 13 25.29337 25.76288 26.20076 26.60912 26.98996 27.34514 27.676372 27.985284
#> 14 24.57082 24.97519 25.35231 25.70402 26.03202 26.33791 26.623191 26.889242
#> 15 22.50171 22.71953 22.92268 23.11213 23.28881 23.45359 23.607262 23.750576
#> 16 17.96937 17.77856 17.60062 17.43466 17.27989 17.13555 17.000941 16.875402
#> 17 24.24239 24.61715 24.96666 25.29261 25.59659 25.88008 26.144472 26.391041
#> 18 25.62836 26.12809 26.59413 27.02876 27.43410 27.81212 28.164665 28.493449
#> 19 20.18628 20.19534 20.20380 20.21168 20.21904 20.22590 20.232293 20.238259
#> 20 21.50000 21.62751 21.74643 21.85733 21.96076 22.05721 22.147169 22.231063
#> 21 18.87256 18.76318 18.66117 18.56604 18.47732 18.39458 18.317418 18.245455
#>       1.6951    1.5808    1.4743    1.3749    1.2823    1.1958    1.1152
#> 1  21.752417 21.752417 21.752417 21.752417 21.752417 21.691748 21.625141
#> 2  24.593222 24.593222 24.593222 24.593222 24.593222 24.635564 24.682049
#> 3  20.012104 20.012104 20.012104 20.012104 20.012104 20.136514 20.273100
#> 4  18.758056 18.758056 18.758056 18.758056 18.758056 18.945353 19.150981
#> 5  18.195013 18.195013 18.195013 18.195013 18.195013 18.064656 17.921542
#> 6  18.860427 18.860427 18.860427 18.860427 18.860427 18.906850 18.957815
#> 7  18.860427 18.860427 18.860427 18.860427 18.860427 18.950724 19.049858
#> 8  15.635730 15.635730 15.635730 15.635730 15.635730 15.631387 15.626620
#> 9  17.376042 17.376042 17.376042 17.376042 17.376042 17.378205 17.380578
#> 10 17.120114 17.120114 17.120114 17.120114 17.120114 17.152720 17.188517
#> 11  9.595821  9.595821  9.595821  9.595821  9.595821  9.662072  9.734808
#> 12  9.109557  9.109557  9.109557  9.109557  9.109557  9.137128  9.167398
#> 13 28.201811 28.201811 28.201811 28.201811 28.201811 28.220735 28.241510
#> 14 27.075727 27.075727 27.075727 27.075727 27.075727 27.200815 27.338145
#> 15 23.851030 23.851030 23.851030 23.851030 23.851030 23.999208 24.161887
#> 16 16.787407 16.787407 16.787407 16.787407 16.787407 16.752098 16.713333
#> 17 26.563870 26.563870 26.563870 26.563870 26.563870 26.618223 26.677895
#> 18 28.723905 28.723905 28.723905 28.723905 28.723905 28.621932 28.509979
#> 19 20.242440 20.242440 20.242440 20.242440 20.242440 20.004544 19.743367
#> 20 22.289867 22.289867 22.289867 22.289867 22.289867 22.115541 21.924156
#> 21 18.195013 18.195013 18.195013 18.195013 18.195013 17.973983 17.731322
#>       1.0401    0.9700    0.9046    0.8436    0.7868    0.7338    0.6843
#> 1  21.563020 21.505086 21.451057 21.400669 21.353677 21.309852 21.268981
#> 2  24.725404 24.765837 24.803544 24.838710 24.871507 24.902092 24.930617
#> 3  20.400485 20.519285 20.630078 20.733404 20.829766 20.919634 21.003445
#> 4  19.342758 19.521610 19.688408 19.843964 19.989036 20.124331 20.250508
#> 5  17.788068 17.663589 17.547500 17.439235 17.338267 17.244103 17.156286
#> 6  19.005348 19.049678 19.091019 19.129575 19.165532 19.199065 19.230338
#> 7  19.142314 19.228539 19.308953 19.383947 19.453887 19.519113 19.579943
#> 8  15.622174 15.618028 15.614161 15.610554 15.607191 15.604054 15.601129
#> 9  17.382792 17.384857 17.386783 17.388578 17.390253 17.391815 17.393272
#> 10 17.221902 17.253038 17.282075 17.309155 17.334410 17.357963 17.379928
#> 11  9.802643  9.865907  9.924908  9.979931 10.031247 10.079104 10.123735
#> 12  9.195628  9.221956  9.246509  9.269408  9.290763  9.310679  9.329253
#> 13 28.260886 28.278955 28.295808 28.311524 28.326181 28.339850 28.352598
#> 14 27.466225 27.585672 27.697069 27.800959 27.897846 27.988204 28.072472
#> 15 24.313609 24.455105 24.587065 24.710132 24.824904 24.931940 25.031763
#> 16 16.677180 16.643462 16.612018 16.582692 16.555343 16.529838 16.506051
#> 17 26.733548 26.785449 26.833853 26.878995 26.921094 26.960356 26.996972
#> 18 28.405567 28.308192 28.217380 28.132689 28.053705 27.980044 27.911348
#> 19 19.499782 19.272613 19.060756 18.863176 18.678913 18.507069 18.346806
#> 20 21.745661 21.579197 21.423952 21.279170 21.144145 21.018221 20.900784
#> 21 17.505006 17.293943 17.097104 16.913532 16.742333 16.582671 16.433770
#>       0.6382   0.5952    0.5551    0.5177    0.4828   0.4502   0.4199   0.3916
#> 1  21.230865 21.19532 21.162165 21.131248 21.102414 22.10429 22.10456 22.10475
#> 2  24.957219 24.98203 25.005165 25.026742 25.046866 26.16586 26.16616 26.16638
#> 3  21.081608 21.15450 21.222484 21.285884 21.345011 20.88220 20.88209 20.88200
#> 4  20.368181 20.47792 20.580268 20.675716 20.764731 20.96797 20.96804 20.96809
#> 5  17.074388 16.99801 16.926778 16.860347 16.798394 15.00962 15.00913 15.00878
#> 6  19.259504 19.28670 19.312071 19.335728 19.357791 18.60756 18.60736 18.60722
#> 7  19.636673 19.68958 19.738921 19.784937 19.827852 19.36822 19.36811 19.36802
#> 8  15.598401 15.59586 15.593484 15.591271 15.589207 15.14430 15.14418 15.14410
#> 9  17.394630 17.39590 17.397078 17.398180 17.399208 16.65114 16.65094 16.65079
#> 10 17.400413 17.41952 17.437335 17.453951 17.469447 16.97394 16.97381 16.97372
#> 11 10.165359 10.20418 10.240379 10.274141 10.305628 11.53000 11.53034 11.53058
#> 12  9.346575  9.36273  9.377795  9.391846  9.404949 10.46986 10.47015 10.47036
#> 13 28.364487 28.37557 28.385915 28.395558 28.404552 28.65047 28.65053 28.65058
#> 14 28.151060 28.22435 28.292705 28.356450 28.415899 29.58906 29.58938 29.58962
#> 15 25.124858 25.21168 25.292649 25.368161 25.438584 24.36942 24.36914 24.36895
#> 16 16.483867 16.46318 16.443885 16.425891 16.409110 15.52996 15.52972 15.52954
#> 17 27.031120 27.06297 27.092666 27.120365 27.146197 27.95266 27.95289 27.95305
#> 18 27.847282 27.78753 27.731812 27.679846 27.631382 26.97265 26.97246 26.97233
#> 19 18.197344 18.05796 17.927962 17.806729 17.693667 17.82208 17.82210 17.82211
#> 20 20.791261 20.68912 20.593863 20.505026 20.422176 20.56431 20.56433 20.56435
#> 21 16.294905 16.16540 16.044620 15.931982 15.826935 16.47442 16.47457 16.47469
#>      0.3652   0.3406   0.3176   0.2962   0.2763   0.2576   0.2403   0.2241
#> 1  22.10488 22.10498 22.10505 22.10510 21.94460 21.94481 21.91494 21.88879
#> 2  26.16654 26.16665 26.16673 26.16678 26.73174 26.73209 26.65604 26.59081
#> 3  20.88195 20.88190 20.88187 20.88185 21.13402 21.13396 20.94478 20.78510
#> 4  20.96813 20.96815 20.96817 20.96819 21.03963 21.03970 20.63032 20.28373
#> 5  15.00853 15.00834 15.00822 15.00812 15.16000 15.15958 15.14843 15.14028
#> 6  18.60711 18.60704 18.60699 18.60695 17.85520 17.85491 18.17220 18.44150
#> 7  19.36797 19.36792 19.36789 19.36787 18.40743 18.40719 18.69064 18.93110
#> 8  15.14404 15.14399 15.14396 15.14394 15.43257 15.43251 15.38879 15.35213
#> 9  16.65069 16.65062 16.65056 16.65052 16.74393 16.74377 16.62612 16.52710
#> 10 16.97365 16.97360 16.97357 16.97355 16.94632 16.94620 16.81520 16.70471
#> 11 11.53075 11.53087 11.53096 11.53102 11.55953 11.55983 11.54389 11.52949
#> 12 10.47051 10.47062 10.47069 10.47075 10.72913 10.72942 10.97010 11.17307
#> 13 28.65062 28.65064 28.65066 28.65067 28.49184 28.49187 28.97795 29.38934
#> 14 29.58978 29.58990 29.58999 29.59005 29.52709 29.52737 29.54513 29.55937
#> 15 24.36881 24.36870 24.36863 24.36858 24.14531 24.14504 24.21438 24.27410
#> 16 15.52942 15.52933 15.52927 15.52923 16.35095 16.35086 16.35246 16.35448
#> 17 27.95316 27.95324 27.95330 27.95334 28.27514 28.27538 28.27113 28.26695
#> 18 26.97223 26.97216 26.97211 26.97208 27.33898 27.33887 27.09509 26.88909
#> 19 17.82212 17.82212 17.82213 17.82213 18.64711 18.64724 19.03523 19.36339
#> 20 20.56436 20.56437 20.56438 20.56438 19.90467 19.90459 19.62717 19.39198
#> 21 16.47477 16.47482 16.47486 16.47489 15.43483 15.43480 15.17999 14.96349
#>      0.2090   0.1949   0.1818   0.1695   0.1581   0.1474   0.1375   0.1282
#> 1  21.88855 21.88838 21.88826 21.88817 21.88811 21.88807 21.88804 22.71328
#> 2  26.59066 26.59056 26.59049 26.59044 26.59041 26.59038 26.59037 26.52511
#> 3  20.78520 20.78528 20.78533 20.78538 20.78541 20.78543 20.78544 20.73554
#> 4  20.28361 20.28354 20.28349 20.28346 20.28344 20.28343 20.28343 20.14959
#> 5  15.14066 15.14093 15.14113 15.14127 15.14137 15.14144 15.14149 14.91972
#> 6  18.44160 18.44166 18.44170 18.44172 18.44173 18.44174 18.44174 18.80279
#> 7  18.93110 18.93109 18.93107 18.93106 18.93104 18.93103 18.93103 19.03006
#> 8  15.35225 15.35235 15.35242 15.35247 15.35250 15.35253 15.35254 15.58559
#> 9  16.52724 16.52734 16.52742 16.52747 16.52751 16.52754 16.52756 16.55604
#> 10 16.70477 16.70482 16.70486 16.70489 16.70491 16.70493 16.70494 16.57598
#> 11 11.52925 11.52907 11.52895 11.52886 11.52880 11.52875 11.52872 11.66022
#> 12 11.17295 11.17287 11.17280 11.17275 11.17271 11.17269 11.17267 11.10196
#> 13 29.38938 29.38939 29.38939 29.38939 29.38938 29.38937 29.38937 29.54053
#> 14 29.55914 29.55896 29.55884 29.55875 29.55869 29.55864 29.55861 29.09263
#> 15 24.27431 24.27445 24.27454 24.27461 24.27466 24.27469 24.27471 23.73886
#> 16 16.35480 16.35503 16.35519 16.35531 16.35539 16.35545 16.35549 16.64394
#> 17 28.26684 28.26676 28.26671 28.26667 28.26665 28.26663 28.26661 28.29723
#> 18 26.88923 26.88933 26.88941 26.88947 26.88952 26.88955 26.88957 27.09834
#> 19 19.36359 19.36372 19.36381 19.36387 19.36391 19.36394 19.36396 18.88610
#> 20 19.39177 19.39162 19.39152 19.39146 19.39141 19.39138 19.39136 20.09901
#> 21 14.96311 14.96285 14.96266 14.96254 14.96245 14.96239 14.96235 14.04748
#>      0.1196   0.1115   0.1040   0.0970   0.0905   0.0844   0.0787   0.0734
#> 1  22.71366 22.71395 22.71417 22.79719 22.79700 22.79684 22.79672 22.79662
#> 2  26.52501 26.52493 26.52486 25.88314 25.88312 25.88310 25.88309 25.88307
#> 3  20.73556 20.73557 20.73558 21.26311 21.26296 21.26282 21.26270 21.26259
#> 4  20.14974 20.14986 20.14996 20.23766 20.23762 20.23759 20.23755 20.23751
#> 5  14.91959 14.91949 14.91942 15.25976 15.25972 15.25969 15.25966 15.25963
#> 6  18.80292 18.80303 18.80310 18.48095 18.48111 18.48127 18.48142 18.48156
#> 7  19.03016 19.03024 19.03030 18.92656 18.92680 18.92702 18.92722 18.92740
#> 8  15.58564 15.58567 15.58569 14.70244 14.70246 14.70248 14.70251 14.70254
#> 9  16.55608 16.55612 16.55615 16.24148 16.24150 16.24151 16.24152 16.24154
#> 10 16.57600 16.57602 16.57604 16.33407 16.33414 16.33420 16.33425 16.33429
#> 11 11.66028 11.66033 11.66037 12.43811 12.43793 12.43777 12.43764 12.43752
#> 12 11.10176 11.10160 11.10147 10.75101 10.75105 10.75108 10.75111 10.75114
#> 13 29.54038 29.54025 29.54014 30.11349 30.11343 30.11339 30.11336 30.11333
#> 14 29.09246 29.09234 29.09224 29.28892 29.28902 29.28910 29.28915 29.28920
#> 15 23.73866 23.73851 23.73840 23.54052 23.54075 23.54094 23.54111 23.54125
#> 16 16.64386 16.64379 16.64372 17.32386 17.32348 17.32316 17.32287 17.32263
#> 17 28.29718 28.29714 28.29710 28.11964 28.11958 28.11953 28.11949 28.11945
#> 18 27.09845 27.09855 27.09862 27.14728 27.14711 27.14697 27.14684 27.14674
#> 19 18.88554 18.88509 18.88473 18.80607 18.80605 18.80602 18.80600 18.80597
#> 20 20.09957 20.10001 20.10036 19.98756 19.98753 19.98751 19.98750 19.98750
#> 21 14.04749 14.04753 14.04757 14.15719 14.15763 14.15799 14.15829 14.15853
#>      0.0684   0.0638   0.0595   0.0555   0.0518   0.0483   0.0450   0.0420
#> 1  22.79654 22.79649 22.79644 22.79641 22.79639 22.79637 22.79636 22.79636
#> 2  25.88306 25.88304 25.88303 25.88302 25.88301 25.88300 25.88299 25.88298
#> 3  21.26250 21.26241 21.26234 21.26228 21.26223 21.26218 21.26214 21.26211
#> 4  20.23747 20.23744 20.23741 20.23738 20.23735 20.23732 20.23730 20.23728
#> 5  15.25961 15.25958 15.25956 15.25954 15.25952 15.25951 15.25949 15.25948
#> 6  18.48169 18.48181 18.48191 18.48201 18.48209 18.48217 18.48223 18.48229
#> 7  18.92755 18.92769 18.92781 18.92792 18.92801 18.92809 18.92815 18.92821
#> 8  14.70256 14.70259 14.70262 14.70264 14.70267 14.70269 14.70271 14.70273
#> 9  16.24155 16.24156 16.24157 16.24158 16.24159 16.24159 16.24160 16.24160
#> 10 16.33432 16.33435 16.33437 16.33439 16.33441 16.33442 16.33443 16.33444
#> 11 12.43741 12.43732 12.43725 12.43718 12.43712 12.43708 12.43704 12.43700
#> 12 10.75116 10.75118 10.75119 10.75121 10.75122 10.75123 10.75124 10.75125
#> 13 30.11332 30.11331 30.11330 30.11330 30.11330 30.11330 30.11331 30.11331
#> 14 29.28923 29.28925 29.28926 29.28927 29.28927 29.28927 29.28927 29.28926
#> 15 23.54136 23.54146 23.54154 23.54160 23.54166 23.54170 23.54173 23.54176
#> 16 17.32243 17.32226 17.32211 17.32199 17.32189 17.32180 17.32173 17.32168
#> 17 28.11942 28.11939 28.11937 28.11935 28.11933 28.11932 28.11930 28.11929
#> 18 27.14664 27.14657 27.14650 27.14645 27.14640 27.14636 27.14633 27.14630
#> 19 18.80594 18.80591 18.80588 18.80586 18.80583 18.80581 18.80579 18.80577
#> 20 19.98751 19.98752 19.98754 19.98755 19.98757 19.98758 19.98760 19.98762
#> 21 14.15872 14.15887 14.15899 14.15909 14.15916 14.15921 14.15925 14.15928
#>      0.0392   0.0365   0.0341   0.0318   0.0296   0.0276   0.0258   0.0240
#> 1  22.79636 22.79636 22.79636 22.79637 22.79637 22.79638 22.79638 22.79620
#> 2  25.88297 25.88297 25.88296 25.88296 25.88295 25.88295 25.88295 25.88281
#> 3  21.26208 21.26206 21.26204 21.26202 21.26201 21.26199 21.26199 21.26182
#> 4  20.23726 20.23725 20.23723 20.23722 20.23721 20.23720 20.23719 20.23722
#> 5  15.25947 15.25946 15.25945 15.25944 15.25944 15.25943 15.25943 15.25868
#> 6  18.48234 18.48239 18.48242 18.48246 18.48249 18.48251 18.48253 18.48331
#> 7  18.92826 18.92830 18.92834 18.92837 18.92839 18.92841 18.92843 18.92928
#> 8  14.70275 14.70276 14.70277 14.70279 14.70280 14.70280 14.70281 14.70255
#> 9  16.24161 16.24161 16.24162 16.24162 16.24162 16.24162 16.24163 16.24128
#> 10 16.33444 16.33445 16.33445 16.33445 16.33445 16.33445 16.33445 16.33417
#> 11 12.43697 12.43695 12.43693 12.43691 12.43690 12.43689 12.43688 12.43698
#> 12 10.75125 10.75126 10.75126 10.75126 10.75127 10.75127 10.75127 10.75129
#> 13 30.11332 30.11332 30.11333 30.11333 30.11334 30.11334 30.11335 30.11302
#> 14 29.28925 29.28925 29.28924 29.28923 29.28922 29.28922 29.28921 29.28911
#> 15 23.54178 23.54180 23.54181 23.54182 23.54182 23.54183 23.54183 23.54152
#> 16 17.32163 17.32160 17.32157 17.32154 17.32153 17.32151 17.32150 17.32111
#> 17 28.11929 28.11928 28.11927 28.11927 28.11926 28.11926 28.11926 28.11908
#> 18 27.14628 27.14626 27.14625 27.14624 27.14623 27.14623 27.14622 27.14663
#> 19 18.80576 18.80574 18.80573 18.80572 18.80571 18.80570 18.80569 18.80601
#> 20 19.98763 19.98765 19.98766 19.98767 19.98768 19.98769 19.98770 19.98824
#> 21 14.15930 14.15931 14.15931 14.15932 14.15931 14.15931 14.15930 14.15966
#>      0.0224   0.0209   0.0195   0.0182   0.0170   0.0158   0.0147   0.0137
#> 1  22.81326 22.81315 22.81303 22.81292 22.81281 22.81271 22.81261 22.81252
#> 2  25.59968 25.59943 25.59920 25.59897 25.59875 25.59854 25.59834 25.59815
#> 3  21.17184 21.17182 21.17181 21.17179 21.17178 21.17177 21.17175 21.17174
#> 4  20.23980 20.23985 20.23991 20.23996 20.24001 20.24005 20.24010 20.24014
#> 5  14.99021 14.99007 14.98992 14.98979 14.98966 14.98953 14.98941 14.98929
#> 6  18.43421 18.43407 18.43392 18.43378 18.43365 18.43353 18.43340 18.43328
#> 7  19.07125 19.07131 19.07136 19.07142 19.07147 19.07151 19.07156 19.07160
#> 8  14.59702 14.59687 14.59673 14.59660 14.59647 14.59634 14.59622 14.59611
#> 9  16.30743 16.30749 16.30754 16.30760 16.30765 16.30770 16.30775 16.30779
#> 10 16.51182 16.51199 16.51216 16.51232 16.51248 16.51263 16.51277 16.51291
#> 11 12.54490 12.54499 12.54509 12.54518 12.54526 12.54534 12.54542 12.54550
#> 12 10.59092 10.59077 10.59062 10.59048 10.59034 10.59021 10.59009 10.58997
#> 13 29.99286 29.99274 29.99263 29.99252 29.99242 29.99232 29.99222 29.99213
#> 14 29.50689 29.50714 29.50739 29.50763 29.50786 29.50809 29.50830 29.50851
#> 15 23.60278 23.60292 23.60306 23.60319 23.60332 23.60344 23.60356 23.60367
#> 16 17.43553 17.43564 17.43575 17.43586 17.43596 17.43606 17.43615 17.43624
#> 17 27.99890 27.99880 27.99869 27.99860 27.99850 27.99841 27.99832 27.99824
#> 18 27.18460 27.18462 27.18463 27.18464 27.18465 27.18466 27.18467 27.18467
#> 19 18.96366 18.96384 18.96401 18.96418 18.96434 18.96449 18.96464 18.96478
#> 20 20.15024 20.15023 20.15023 20.15023 20.15022 20.15022 20.15022 20.15022
#> 21 14.09220 14.09225 14.09230 14.09235 14.09240 14.09244 14.09248 14.09252
#>      0.0128   0.0120   0.0112   0.0104   0.0097   0.0090   0.0084   0.0079
#> 1  22.81243 22.81234 22.81226 22.81218 22.81210 22.81202 22.81183 22.80606
#> 2  25.59797 25.59779 25.59761 25.59745 25.59729 25.59714 25.59713 25.61027
#> 3  21.17173 21.17172 21.17171 21.17170 21.17169 21.17168 21.17172 21.14061
#> 4  20.24018 20.24022 20.24026 20.24030 20.24033 20.24037 20.24036 20.19950
#> 5  14.98918 14.98907 14.98897 14.98887 14.98878 14.98868 14.98864 15.00649
#> 6  18.43317 18.43306 18.43296 18.43286 18.43276 18.43267 18.43241 18.41076
#> 7  19.07165 19.07169 19.07173 19.07177 19.07180 19.07184 19.07170 19.05084
#> 8  14.59600 14.59589 14.59579 14.59569 14.59560 14.59550 14.59516 14.56297
#> 9  16.30784 16.30788 16.30792 16.30796 16.30800 16.30804 16.30782 16.27339
#> 10 16.51305 16.51317 16.51330 16.51342 16.51353 16.51364 16.51349 16.47996
#> 11 12.54557 12.54564 12.54571 12.54578 12.54584 12.54590 12.54625 12.59029
#> 12 10.58985 10.58974 10.58964 10.58953 10.58943 10.58934 10.58945 10.63760
#> 13 29.99204 29.99196 29.99187 29.99179 29.99172 29.99164 29.99157 29.97333
#> 14 29.50870 29.50889 29.50907 29.50925 29.50942 29.50958 29.50974 29.49802
#> 15 23.60378 23.60388 23.60398 23.60408 23.60417 23.60426 23.60445 23.70735
#> 16 17.43632 17.43641 17.43649 17.43656 17.43664 17.43671 17.43687 17.43489
#> 17 27.99816 27.99808 27.99801 27.99793 27.99787 27.99780 27.99777 27.98777
#> 18 27.18468 27.18469 27.18470 27.18471 27.18471 27.18472 27.18492 27.22166
#> 19 18.96492 18.96505 18.96518 18.96530 18.96542 18.96553 18.96576 18.93248
#> 20 20.15021 20.15021 20.15021 20.15021 20.15020 20.15020 20.15015 20.19430
#> 21 14.09256 14.09260 14.09264 14.09267 14.09270 14.09274 14.09280 14.08148
#>      0.0073   0.0068   0.0064   0.0060   0.0056   0.0052
#> 1  22.80615 22.80621 22.80627 22.80631 22.80633 22.78779
#> 2  25.60999 25.60972 25.60945 25.60920 25.60895 25.62513
#> 3  21.14040 21.14022 21.14007 21.13994 21.13984 21.12701
#> 4  20.19934 20.19921 20.19911 20.19902 20.19896 20.19601
#> 5  15.00619 15.00591 15.00564 15.00538 15.00514 14.98832
#> 6  18.41104 18.41130 18.41152 18.41171 18.41188 18.40280
#> 7  19.05123 19.05159 19.05191 19.05221 19.05248 19.04460
#> 8  14.56292 14.56286 14.56280 14.56273 14.56266 14.57876
#> 9  16.27339 16.27339 16.27339 16.27340 16.27341 16.28003
#> 10 16.48002 16.48008 16.48015 16.48022 16.48030 16.48917
#> 11 12.59038 12.59047 12.59057 12.59066 12.59075 12.57928
#> 12 10.63757 10.63753 10.63748 10.63741 10.63734 10.65009
#> 13 29.97326 29.97319 29.97312 29.97305 29.97298 29.94770
#> 14 29.49804 29.49808 29.49813 29.49818 29.49824 29.50458
#> 15 23.70765 23.70791 23.70813 23.70832 23.70849 23.72333
#> 16 17.43490 17.43493 17.43498 17.43504 17.43512 17.43814
#> 17 27.98760 27.98745 27.98730 27.98717 27.98704 27.99092
#> 18 27.22174 27.22183 27.22192 27.22201 27.22210 27.23400
#> 19 18.93239 18.93232 18.93229 18.93228 18.93229 18.96817
#> 20 20.19465 20.19497 20.19525 20.19550 20.19571 20.18540
#> 21 14.08114 14.08082 14.08053 14.08026 14.08001 14.05876
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.1 14.3 19.2 17.8 16.4 17.3 15.2 10.4 14.7 30.4 33.9 21.5
#> [16] 19.2 27.3 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.1547005  0.5557189 -0.1714986 -0.57162624  0.60023193  0.3735437
#>  [2,]  1.1547005 -1.1114379 -1.3719887 -0.99811972  0.50658505  0.3735437
#>  [3,] -0.8660254 -1.1114379 -0.1714986  0.23214993 -0.93557696 -0.9338592
#>  [4,] -0.8660254 -1.1114379 -0.1714986 -0.03850939 -1.53491702 -0.9338592
#>  [5,] -0.8660254  0.5557189  1.0289915  1.06873329 -0.69209507 -0.9338592
#>  [6,] -0.8660254  0.5557189 -0.1714986 -0.50929258  0.63769069  0.3735437
#>  [7,] -0.8660254  0.5557189 -0.1714986 -0.50929258  0.63769069  0.3735437
#>  [8,] -0.8660254  0.0000000  1.0289915  0.37814193 -0.95430634 -0.9338592
#>  [9,] -0.8660254  0.0000000  1.0289915  0.37814193 -0.95430634 -0.9338592
#> [10,] -0.8660254  0.0000000  1.0289915  0.37814193 -0.95430634 -0.9338592
#> [11,] -0.8660254  0.5557189  1.0289915  1.98733463 -1.21651762 -0.9338592
#> [12,] -0.8660254  0.5557189  1.0289915  1.72487710 -0.65463631 -0.9338592
#> [13,]  1.1547005 -0.5557189 -1.3719887 -1.26303778  2.52935774  0.3735437
#> [14,]  1.1547005 -1.1114379 -1.3719887 -1.30076605  1.19957199  0.3735437
#> [15,] -0.8660254 -1.1114379 -1.3719887 -0.89887797  0.22564440 -0.9338592
#> [16,] -0.8660254 -0.5557189  1.0289915  1.39680520 -0.93557696 -0.9338592
#> [17,]  1.1547005 -1.1114379 -1.3719887 -1.23597185  0.93736071  0.3735437
#> [18,]  1.1547005 -0.5557189 -1.3719887 -1.10392291  0.35675003  1.6809466
#> [19,]  1.1547005  0.5557189  1.0289915  0.99491711  1.19957199  1.6809466
#> [20,]  1.1547005  1.6671568 -0.1714986 -0.69465320  0.07580938  1.6809466
#> [21,]  1.1547005  2.7785947  1.0289915  0.58482723 -0.07402563  1.6809466
#>               hp       qsec         vs          wt
#>  [1,] -0.6101029 -0.4499273 -0.9534626 -0.32197332
#>  [2,] -0.8478832  0.5390834  1.0488088 -0.87075939
#>  [3,] -0.6101029  1.0553594  1.0488088  0.01421994
#>  [4,] -0.6800383  1.5405345  1.0488088  0.25647685
#>  [5,]  1.2781523 -1.1839101 -0.9534626  0.36524526
#>  [6,] -0.4282709  0.3462574  1.0488088  0.23670078
#>  [7,] -0.4282709  0.7194690  1.0488088  0.23670078
#>  [8,]  0.3689924 -0.2135600 -0.9534626  0.85964712
#>  [9,]  0.3689924 -0.0891561 -0.9534626  0.52345385
#> [10,]  0.3689924  0.1596516 -0.9534626  0.57289404
#> [11,]  0.7186693  0.1472112 -0.9534626  2.02643550
#> [12,]  1.0683462 -0.2011196 -0.9534626  2.12037186
#> [13,] -1.4213533  0.4831017  1.0488088 -1.56786601
#> [14,] -1.2395213  1.3414883  1.0488088 -1.35032919
#> [15,] -0.7919349  1.4099104  1.0488088 -0.72738285
#> [16,]  0.2990570 -0.4312667 -0.9534626  0.63716628
#> [17,] -1.2255343  0.7194690  1.0488088 -1.25144882
#> [18,] -0.5681417 -0.5245696  1.0488088 -1.66872399
#> [19,]  1.5439067 -2.0174159 -0.9534626 -0.03027623
#> [20,]  0.2990570 -1.3953966 -0.9534626 -0.42579771
#> [21,]  2.5369891 -1.9552140 -0.9534626  0.36524526
#> attr(,"center")
#>  [1]   0.4285714   3.0000000   6.2857143 229.6952381   3.5795238   3.7142857
#>  [7] 153.6190476  17.7433333   0.4761905   3.2006190
#> attr(,"scale")
#>  [1]   0.4948717   1.7994708   1.6659863 121.9244906   0.5339206   0.7648752
#>  [7]  71.4945727   1.6076671   0.4994328   1.0113231
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 27
#> 
#> $lambda.min
#> [1] 0.8436445
#> 
#> $null.dev
#> [1] 34.48599
#> 
#> $Bias
#> [1] 8.10223
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.917073