Regression MARS (Multivariate Adaptive Regression Splines) Learner

Calls earth::earth from package earth.

Details

Methods for variance estimations are not yet implemented.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.earth")
lrn("regr.earth")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “numeric”, “factor”, “integer”

Required Packages: mlr3extralearners, earth

Parameters

Id	Type	Default	Levels	Range

wp	list	NULL		\((-\infty, \infty)\)

offset	list	NULL		\((-\infty, \infty)\)

keepxy	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

trace	character	0	0, .3, .5, 1, 2, 3, 4, 5	\((-\infty, \infty)\)

degree	integer	1		\([1, \infty)\)

penalty	numeric	2		\([-1, \infty)\)

nk	list	NULL		\((-\infty, \infty)\)

thresh	numeric	0.001		\((-\infty, \infty)\)

minspan	numeric	0		\([0, \infty)\)

endspan	numeric	0		\([0, \infty)\)

newvar.penalty	numeric	0		\([0, \infty)\)

fast.k	integer	20		\([0, \infty)\)

fast.beta	integer	1		\([0, 1]\)

linpreds	list	FALSE		\((-\infty, \infty)\)

allowed	list	-		\((-\infty, \infty)\)

pmethod	character	backward	backward, none, exhaustive, forward, seqrep, cv	\((-\infty, \infty)\)

nprune	integer	-		\([0, \infty)\)

nfold	integer	0		\([0, \infty)\)

ncross	integer	1		\([0, \infty)\)

stratify	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

varmod.method	character	none	none, const, lm, rlm, earth, gam, power, power0, x.lm, x.rlm, ...	\((-\infty, \infty)\)

varmod.exponent	numeric	1		\((-\infty, \infty)\)

varmod.conv	numeric	1		\([0, 1]\)

varmod.clamp	numeric	0.1		\((-\infty, \infty)\)

varmod.minspan	numeric	-3		\((-\infty, \infty)\)

Scale.y	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

Adjust.endspan	numeric	2		\((-\infty, \infty)\)

Auto.linpreds	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

Force.weights	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

Use.beta.cache	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

Force.xtx.prune	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

Get.leverages	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

Exhaustive.tol	numeric	1e-10		\((-\infty, \infty)\)

References

Stephen Milborrow (2014) Earth: multivariate adaptive regression spline models R package version 3 https://cran.r-project.org/web/packages/earth/earth.pdf

Jerome H. Friedman (1991) Multivariate Adaptive Regression Splines The Annals of Statistics https://projecteuclid.org/download/pdf_1/euclid.aos/1176347963

See also

Author

pkopper

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrEarth

Methods

Public methods

LearnerRegrEarth$new()

LearnerRegrEarth$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrEarth$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrEarth$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("earth", quietly = TRUE)) {
  learner = mlr3::lrn("regr.earth")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerRegrEarth:regr.earth>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3extralearners, earth
#> * Predict Type: response
#> * Feature types: numeric, factor, integer
#> * Properties: weights
#>  [1] "wp"              "offset"          "keepxy"          "trace"          
#>  [5] "degree"          "penalty"         "nk"              "thresh"         
#>  [9] "minspan"         "endspan"         "newvar.penalty"  "fast.k"         
#> [13] "fast.beta"       "linpreds"        "allowed"         "pmethod"        
#> [17] "nprune"          "nfold"           "ncross"          "stratify"       
#> [21] "varmod.method"   "varmod.exponent" "varmod.conv"     "varmod.clamp"   
#> [25] "varmod.minspan"  "Scale.y"         "Adjust.endspan"  "Auto.linpreds"  
#> [29] "Force.weights"   "Use.beta.cache"  "Force.xtx.prune" "Get.leverages"  
#> [33] "Exhaustive.tol"