Boosted Generalized Additive Regression Learner

Fit a generalized additive regression model using a boosting algorithm. Calls mboost::gamboost() from mboost.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.gamboost")
lrn("regr.gamboost")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, mboost

Parameters

Id	Type	Default	Levels	Range

baselearner	character	bbs	bbs, bols, btree	-

dfbase	integer	4		\((-\infty, \infty)\)

offset	numeric	NULL		\((-\infty, \infty)\)

family	character	Gaussian	Gaussian, Laplace, Huber, Poisson, GammaReg, NBinomial, Hurdle, custom	-

custom.family	untyped	-		-

nuirange	untyped	c(0, 100)		-

d	numeric	NULL		\((-\infty, \infty)\)

mstop	integer	100		\((-\infty, \infty)\)

nu	numeric	0.1		\((-\infty, \infty)\)

risk	character	inbag	inbag, oobag, none	-

oobweights	untyped	NULL		-

trace	logical	FALSE	TRUE, FALSE	-

stopintern	untyped	FALSE		-

na.action	untyped	stats::na.omit		-

References

Bühlmann, Peter, Yu, Bin (2003). “Boosting with the L 2 loss: regression and classification.” Journal of the American Statistical Association, 98(462), 324--339.

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrGAMBoost

Methods

Public methods

LearnerRegrGAMBoost$new()

LearnerRegrGAMBoost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Create a LearnerRegrGAMBoost object.

Usage

LearnerRegrGAMBoost$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrGAMBoost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("regr.gamboost")
print(learner)
#> <LearnerRegrGAMBoost:regr.gamboost>: Boosted Generalized Additive Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, mboost
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: weights

# available parameters:
learner$param_set$ids()
#>  [1] "baselearner"   "dfbase"        "offset"        "family"       
#>  [5] "custom.family" "nuirange"      "d"             "mstop"        
#>  [9] "nu"            "risk"          "oobweights"    "trace"        
#> [13] "stopintern"    "na.action"