Regression Gaussian Process Learner

Calls kernlab::gausspr from package kernlab.

Details

Parameters sigma, degree, scale, offset and order are added to make tuning kpar easier. If kpar is provided then these new parameters are ignored. If none are provided then the default "automatic" is used for kpar.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.gausspr")
lrn("regr.gausspr")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “numeric”, “integer”, “logical”, “character”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, kernlab

Parameters

Id	Type	Default	Levels	Range

scaled	list	TRUE		-

kernel	character	rbfdot	rbfdot, polydot, vanilladot, tanhdot, laplacedot, besseldot, anovadot, splinedot	-

sigma	numeric	-		\((-\infty, \infty)\)

degree	numeric	-		\((-\infty, \infty)\)

scale	numeric	-		\((-\infty, \infty)\)

offset	numeric	-		\((-\infty, \infty)\)

order	numeric	-		\((-\infty, \infty)\)

kpar	list	automatic		-

var	numeric	0.001		\([0.001, \infty)\)

variance.model	logical	FALSE	TRUE, FALSE	-

tol	numeric	0.001		\([0, \infty)\)

fit	logical	TRUE	TRUE, FALSE	-

cross	integer	0		\([0, \infty)\)

na.action	list	function (object, ...) , UseMethod("na.omit")		-

type	character	regression	regression, variance, sdeviation	-

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrGausspr

Methods

Public methods

LearnerRegrGausspr$new()

LearnerRegrGausspr$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrGausspr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrGausspr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("kernlab", quietly = TRUE)) {
  learner = mlr3::lrn("regr.gausspr")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerRegrGausspr:regr.gausspr>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, kernlab
#> * Predict Type: response
#> * Feature types: numeric, integer, logical, character, factor, ordered
#> * Properties: -
#>  [1] "scaled"         "kernel"         "sigma"          "degree"        
#>  [5] "scale"          "offset"         "order"          "kpar"          
#>  [9] "var"            "variance.model" "tol"            "fit"           
#> [13] "na.action"