L2-Regularized Support Vector Regression Learner

L2 regularized support vector regression. Calls LiblineaR::LiblineaR() from LiblineaR.

Details

Type of SVR depends on type argument:

type = 11 - L2-regularized L2-loss support vector regression (primal)

type = 12 – L2-regularized L2-loss support vector regression (dual)

type = 13 – L2-regularized L1-loss support vector regression (dual)

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.liblinear")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, LiblineaR

Parameters

Id	Type	Default	Levels	Range

type	integer	11		\([11, 13]\)

cost	numeric	1		\([0, \infty)\)

bias	numeric	1		\((-\infty, \infty)\)

svr_eps	numeric	NULL		\([0, \infty)\)

cross	integer	0		\([0, \infty)\)

verbose	logical	FALSE	TRUE, FALSE	-

findC	logical	FALSE	TRUE, FALSE	-

useInitC	logical	TRUE	TRUE, FALSE	-

Initial parameter values

svr_eps:

Actual default: NULL

Initial value: 0.001

Reason for change: svr_eps is type dependent and the "type" is handled by the mlr3learner. The default value is set to th default of the respective "type".

References

Fan, Rong-En, Chang, Kai-Wei, Hsieh, Cho-Jui, Wang, Xiang-Rui, Lin, Chih-Jen (2008). “LIBLINEAR: A library for large linear classification.” the Journal of machine Learning research, 9, 1871–1874.

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrLiblineaR

Methods

Public methods

LearnerRegrLiblineaR$new()

LearnerRegrLiblineaR$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrLiblineaR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrLiblineaR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.liblinear")
print(learner)
#> 
#> ── <LearnerRegrLiblineaR> (regr.liblinear): Support Vector Machine ─────────────
#> • Model: -
#> • Parameters: svr_eps=0.001
#> • Packages: mlr3, mlr3extralearners, and LiblineaR
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties:
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $TypeDetail
#> [1] "L2-regularized L2-loss support vector regression primal (L2R_L2LOSS_SVR)"
#> 
#> $Type
#> [1] 11
#> 
#> $W
#>              am       carb       cyl        disp      drat      gear
#> [1,] 0.06866125 0.01036496 0.1925443 -0.03576865 0.3007654 0.2859522
#>               hp     qsec         vs         wt       Bias
#> [1,] 0.002335054 1.397018 0.08934442 0.09080529 0.06987336
#> 
#> $Bias
#> [1] 1
#> 
#> $NbClass
#> [1] 2
#> 
#> attr(,"class")
#> [1] "LiblineaR"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 16.15237