Regression Light GBM Learner

Calls lightgbm::lgb.train from package lightgbm.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.lightgbm")
lrn("regr.lightgbm")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “numeric”, “integer”

Required Packages: mlr3extralearners, lightgbm

Parameters

Id	Type	Default	Levels	Range

nrounds	integer	5		\([1, \infty)\)

objective	character	regression	regression, regression_l1, huber, fair, poisson, quantile, mape, gamma, tweedie	\((-\infty, \infty)\)

metric	character		, None, l1, l2, rmse, quantile, mape, huber, fair, poisson, ...	\((-\infty, \infty)\)

custom_eval	list	NULL		\((-\infty, \infty)\)

verbose	integer	1		\((-\infty, \infty)\)

record	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

eval_freq	integer	1		\([1, \infty)\)

init_model	list	NULL		\((-\infty, \infty)\)

early_stopping_rounds	integer	-		\([1, \infty)\)

callbacks	list	-		\((-\infty, \infty)\)

reset_data	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

categorical_feature	list			\((-\infty, \infty)\)

boosting	character	gbdt	gbdt, rf, dart, goss	\((-\infty, \infty)\)

linear_tree	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_iterations	integer	100		\([0, \infty)\)

learning_rate	numeric	0.1		\([0, \infty)\)

num_leaves	integer	31		\([1, 131072]\)

tree_learner	character	serial	serial, feature, data, voting	\((-\infty, \infty)\)

num_threads	integer	0		\([0, \infty)\)

device_type	character	cpu	cpu, gpu	\((-\infty, \infty)\)

seed	integer	-		\((-\infty, \infty)\)

deterministic	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

force_col_wise	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

force_row_wise	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

histogram_pool_size	integer	-1		\((-\infty, \infty)\)

max_depth	integer	-1		\((-\infty, \infty)\)

min_data_in_leaf	integer	20		\([0, \infty)\)

min_sum_hessian_in_leaf	numeric	0.001		\([0, \infty)\)

bagging_fraction	numeric	1		\([0, 1]\)

bagging_freq	integer	0		\([0, \infty)\)

bagging_seed	integer	3		\((-\infty, \infty)\)

feature_fraction	numeric	1		\([0, 1]\)

feature_fraction_bynode	numeric	1		\([0, 1]\)

feature_fraction_seed	integer	2		\((-\infty, \infty)\)

extra_trees	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

extra_seed	integer	6		\((-\infty, \infty)\)

first_metric_only	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

max_delta_step	numeric	0		\((-\infty, \infty)\)

lambda_l1	numeric	0		\([0, \infty)\)

lambda_l2	numeric	0		\([0, \infty)\)

linear_lambda	numeric	0		\([0, \infty)\)

min_gain_to_split	numeric	0		\([0, \infty)\)

drop_rate	numeric	0.1		\([0, 1]\)

max_drop	integer	50		\((-\infty, \infty)\)

skip_drop	numeric	0.5		\([0, 1]\)

xgboost_dart_mode	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

uniform_drop	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

drop_seed	integer	4		\((-\infty, \infty)\)

top_rate	numeric	0.2		\([0, 1]\)

other_rate	numeric	0.1		\([0, 1]\)

min_data_per_group	integer	100		\([1, \infty)\)

max_cat_threshold	integer	32		\([1, \infty)\)

cat_l2	numeric	10		\([0, \infty)\)

cat_smooth	numeric	10		\([0, \infty)\)

max_cat_to_onehot	integer	4		\([1, \infty)\)

top_k	integer	20		\([1, \infty)\)

monotone_constraints	list	NULL		\((-\infty, \infty)\)

monotone_constraints_method	character	basic	basic, intermediate, advanced	\((-\infty, \infty)\)

monotone_penalty	numeric	0		\([0, \infty)\)

feature_contri	list	NULL		\((-\infty, \infty)\)

forcedsplits_filename	list			\((-\infty, \infty)\)

refit_decay_rate	numeric	0.9		\([0, 1]\)

cegb_tradeoff	numeric	1		\([0, \infty)\)

cegb_penalty_split	numeric	0		\([0, \infty)\)

cegb_penalty_feature_lazy	list	-		\((-\infty, \infty)\)

cegb_penalty_feature_coupled	list	-		\((-\infty, \infty)\)

path_smooth	numeric	0		\([0, \infty)\)

interaction_constraints	list	-		\((-\infty, \infty)\)

input_model	list			\((-\infty, \infty)\)

output_model	list	LightGBM_model.txt		\((-\infty, \infty)\)

saved_feature_importance_type	integer	0		\([0, 1]\)

snapshot_freq	integer	-1		\((-\infty, \infty)\)

max_bin	integer	255		\([2, \infty)\)

max_bin_by_feature	list	NULL		\((-\infty, \infty)\)

min_data_in_bin	integer	3		\([1, \infty)\)

bin_construct_sample_cnt	integer	200000		\([1, \infty)\)

data_random_seed	integer	1		\((-\infty, \infty)\)

is_enable_sparse	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

enable_bundle	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

use_missing	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

zero_as_missing	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

feature_pre_filter	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

pre_partition	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

two_round	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

header	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

group_column	list			\((-\infty, \infty)\)

forcedbins_filename	list			\((-\infty, \infty)\)

save_binary	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

boost_from_average	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

reg_sqrt	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

alpha	numeric	0.9		\([0, \infty)\)

fair_c	numeric	1		\([0, \infty)\)

poisson_max_delta_step	numeric	0.7		\([0, \infty)\)

tweedie_variance_power	numeric	1.5		\([1, 2]\)

metric_freq	integer	1		\([1, \infty)\)

is_provide_training_metric	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_machines	integer	1		\([1, \infty)\)

local_listen_port	integer	12400		\([1, \infty)\)

time_out	integer	120		\([1, \infty)\)

machine_list_filename	list			\((-\infty, \infty)\)

machines	list			\((-\infty, \infty)\)

gpu_platform_id	integer	-1		\((-\infty, \infty)\)

gpu_device_id	integer	-1		\((-\infty, \infty)\)

gpu_use_dp	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_gpu	integer	1		\([1, \infty)\)

start_iteration	integer	0		\((-\infty, \infty)\)

num_iteration	integer	-1		\((-\infty, \infty)\)

pred_early_stop	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

pred_early_stop_freq	integer	10		\((-\infty, \infty)\)

pred_early_stop_margin	numeric	10		\((-\infty, \infty)\)

output_result	list	LightGBM_predict_result.txt		\((-\infty, \infty)\)

Custom mlr3 defaults

num_threads:

Actual default: 0L

Adjusted default: 1L

Reason for change: Prevents accidental conflicts with future.

verbose:

Actual default: 1L

Adjusted default: -1L

Reason for change: Prevents accidental conflicts with mlr messaging system.

For categorical features either pre-process data by encoding columns or specify the categorical columns with the categorical_feature parameter. For this learner please do not prefix the categorical feature with name:.

See also

Author

kapsner

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrLightGBM

Methods

Public methods

LearnerRegrLightGBM$new()

LearnerRegrLightGBM$importance()

LearnerRegrLightGBM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrLightGBM$new()

Method importance()

The importance scores are extracted from lbg.importance.

Usage

LearnerRegrLightGBM$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrLightGBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("lightgbm", quietly = TRUE)) {
  learner = mlr3::lrn("regr.lightgbm")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerRegrLightGBM:regr.lightgbm>
#> * Model: -
#> * Parameters: num_threads=1, verbose=-1, objective=regression
#> * Packages: mlr3, mlr3extralearners, lightgbm
#> * Predict Type: response
#> * Feature types: numeric, integer
#> * Properties: importance, missings, weights
#>   [1] "nrounds"                       "objective"                    
#>   [3] "metric"                        "custom_eval"                  
#>   [5] "verbose"                       "record"                       
#>   [7] "eval_freq"                     "init_model"                   
#>   [9] "early_stopping_rounds"         "callbacks"                    
#>  [11] "reset_data"                    "categorical_feature"          
#>  [13] "boosting"                      "linear_tree"                  
#>  [15] "num_iterations"                "learning_rate"                
#>  [17] "num_leaves"                    "tree_learner"                 
#>  [19] "num_threads"                   "device_type"                  
#>  [21] "seed"                          "deterministic"                
#>  [23] "force_col_wise"                "force_row_wise"               
#>  [25] "histogram_pool_size"           "max_depth"                    
#>  [27] "min_data_in_leaf"              "min_sum_hessian_in_leaf"      
#>  [29] "bagging_fraction"              "bagging_freq"                 
#>  [31] "bagging_seed"                  "feature_fraction"             
#>  [33] "feature_fraction_bynode"       "feature_fraction_seed"        
#>  [35] "extra_trees"                   "extra_seed"                   
#>  [37] "first_metric_only"             "max_delta_step"               
#>  [39] "lambda_l1"                     "lambda_l2"                    
#>  [41] "linear_lambda"                 "min_gain_to_split"            
#>  [43] "drop_rate"                     "max_drop"                     
#>  [45] "skip_drop"                     "xgboost_dart_mode"            
#>  [47] "uniform_drop"                  "drop_seed"                    
#>  [49] "top_rate"                      "other_rate"                   
#>  [51] "min_data_per_group"            "max_cat_threshold"            
#>  [53] "cat_l2"                        "cat_smooth"                   
#>  [55] "max_cat_to_onehot"             "top_k"                        
#>  [57] "monotone_constraints"          "monotone_constraints_method"  
#>  [59] "monotone_penalty"              "feature_contri"               
#>  [61] "forcedsplits_filename"         "refit_decay_rate"             
#>  [63] "cegb_tradeoff"                 "cegb_penalty_split"           
#>  [65] "cegb_penalty_feature_lazy"     "cegb_penalty_feature_coupled" 
#>  [67] "path_smooth"                   "interaction_constraints"      
#>  [69] "input_model"                   "output_model"                 
#>  [71] "saved_feature_importance_type" "snapshot_freq"                
#>  [73] "max_bin"                       "max_bin_by_feature"           
#>  [75] "min_data_in_bin"               "bin_construct_sample_cnt"     
#>  [77] "data_random_seed"              "is_enable_sparse"             
#>  [79] "enable_bundle"                 "use_missing"                  
#>  [81] "zero_as_missing"               "feature_pre_filter"           
#>  [83] "pre_partition"                 "two_round"                    
#>  [85] "header"                        "group_column"                 
#>  [87] "forcedbins_filename"           "save_binary"                  
#>  [89] "boost_from_average"            "reg_sqrt"                     
#>  [91] "alpha"                         "fair_c"                       
#>  [93] "poisson_max_delta_step"        "tweedie_variance_power"       
#>  [95] "metric_freq"                   "is_provide_training_metric"   
#>  [97] "num_machines"                  "local_listen_port"            
#>  [99] "time_out"                      "machine_list_filename"        
#> [101] "machines"                      "gpu_platform_id"              
#> [103] "gpu_device_id"                 "gpu_use_dp"                   
#> [105] "num_gpu"                       "start_iteration"              
#> [107] "num_iteration"                 "pred_early_stop"              
#> [109] "pred_early_stop_freq"          "pred_early_stop_margin"       
#> [111] "output_result"