Regression M5P Learner

Implements base routines for generating M5 Model trees and rules. Calls RWeka::M5P() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.m5p")
lrn("regr.m5p")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

N	logical	-	TRUE, FALSE	-

U	logical	-	TRUE, FALSE	-

R	logical	-	TRUE, FALSE	-

M	integer	4		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

L	logical	-	TRUE, FALSE	-

options	untyped	NULL		-

References

Quinlan RJ (1992). “Learning with Continuous Classes.” In 5th Australian Joint Conference on Artificial Intelligence, 343-348.

Wang Y, Witten IH (1997). “Induction of model trees for predicting continuous classes.” In Poster papers of the 9th European Conference on Machine Learning.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrM5P

Methods

Public methods

LearnerRegrM5P$new()

LearnerRegrM5P$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrM5P$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrM5P$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("regr.m5p")
print(learner)
#> <LearnerRegrM5P:regr.m5p>: M5P
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response]
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: -

# available parameters:
learner$param_set$ids()
#>  [1] "subset"                    "na.action"                
#>  [3] "N"                         "U"                        
#>  [5] "R"                         "M"                        
#>  [7] "output_debug_info"         "do_not_check_capabilities"
#>  [9] "num_decimal_places"        "batch_size"               
#> [11] "L"                         "options"