Regression Mars Learner

Multivariate Adaptive Regression Splines. Calls mda::mars() from mda.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.mars")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, mda

Parameters

Id	Type	Default	Levels	Range

degree	integer	1		\([1, \infty)\)

nk	integer	-		\([1, \infty)\)

penalty	numeric	2		\([0, \infty)\)

thresh	numeric	0.001		\([0, \infty)\)

prune	logical	TRUE	TRUE, FALSE	-

trace.mars	logical	FALSE	TRUE, FALSE	-

forward.step	logical	FALSE	TRUE, FALSE	-

References

Friedman, H J (1991). “Multivariate adaptive regression splines.” The annals of statistics, 19(1), 1–67.

See also

Author

sumny

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrMars

Methods

Public methods

LearnerRegrMars$new()

LearnerRegrMars$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrMars$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrMars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("regr.mars")
print(learner)
#> <LearnerRegrMars:regr.mars>: Multivariate Adaptive Regression Splines
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, mda
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric
#> * Properties: -

# Define a Task
task = mlr3::tsk("mtcars")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $call
#> mda::mars(x = x, y = y)
#> 
#> $all.terms
#>  [1]  1  2  4  6  8 10 12 14 16 18
#> 
#> $selected.terms
#> [1] 1 2 4 6
#> 
#> $penalty
#> [1] 2
#> 
#> $degree
#> [1] 1
#> 
#> $nk
#> [1] 21
#> 
#> $thresh
#> [1] 0.001
#> 
#> $gcv
#> [1] 11.8148
#> 
#> $factor
#>       am carb cyl disp drat gear hp qsec vs wt
#>  [1,]  0    0   0    0    0    0  0    0  0  0
#>  [2,]  0    0   0    0    0    0  0    0  0  1
#>  [3,]  0    0   0    0    0    0  0    0  0 -1
#>  [4,]  0    0   0    0    0    0  0    1  0  0
#>  [5,]  0    0   0    0    0    0  0   -1  0  0
#>  [6,]  1    0   0    0    0    0  0    0  0  0
#>  [7,] -1    0   0    0    0    0  0    0  0  0
#>  [8,]  0    0   1    0    0    0  0    0  0  0
#>  [9,]  0    0  -1    0    0    0  0    0  0  0
#> [10,]  0    0   0    0    1    0  0    0  0  0
#> [11,]  0    0   0    0   -1    0  0    0  0  0
#> [12,]  0    0   0    1    0    0  0    0  0  0
#> [13,]  0    0   0   -1    0    0  0    0  0  0
#> [14,]  0    0   0    0    0    1  0    0  0  0
#> [15,]  0    0   0    0    0   -1  0    0  0  0
#> [16,]  0    1   0    0    0    0  0    0  0  0
#> [17,]  0   -1   0    0    0    0  0    0  0  0
#> [18,]  0    0   0    0    0    0  0    0  1  0
#> [19,]  0    0   0    0    0    0  0    0 -1  0
#> 
#> $cuts
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0  0.0 0.00    0    0  0.0    0 0.000
#>  [2,]    0    0    0  0.0 0.00    0    0  0.0    0 1.615
#>  [3,]    0    0    0  0.0 0.00    0    0  0.0    0 1.615
#>  [4,]    0    0    0  0.0 0.00    0    0 14.6    0 0.000
#>  [5,]    0    0    0  0.0 0.00    0    0 14.6    0 0.000
#>  [6,]    0    0    0  0.0 0.00    0    0  0.0    0 0.000
#>  [7,]    0    0    0  0.0 0.00    0    0  0.0    0 0.000
#>  [8,]    0    0    4  0.0 0.00    0    0  0.0    0 0.000
#>  [9,]    0    0    4  0.0 0.00    0    0  0.0    0 0.000
#> [10,]    0    0    0  0.0 2.76    0    0  0.0    0 0.000
#> [11,]    0    0    0  0.0 2.76    0    0  0.0    0 0.000
#> [12,]    0    0    0 71.1 0.00    0    0  0.0    0 0.000
#> [13,]    0    0    0 71.1 0.00    0    0  0.0    0 0.000
#> [14,]    0    0    0  0.0 0.00    3    0  0.0    0 0.000
#> [15,]    0    0    0  0.0 0.00    3    0  0.0    0 0.000
#> [16,]    0    1    0  0.0 0.00    0    0  0.0    0 0.000
#> [17,]    0    1    0  0.0 0.00    0    0  0.0    0 0.000
#> [18,]    0    0    0  0.0 0.00    0    0  0.0    0 0.000
#> [19,]    0    0    0  0.0 0.00    0    0  0.0    0 0.000
#> 
#> $residuals
#>              [,1]
#>  [1,] -0.90273937
#>  [2,] -3.86001855
#>  [3,] -0.02287039
#>  [4,]  2.55171738
#>  [5,]  1.84243104
#>  [6,]  0.63940627
#>  [7,] -1.89136456
#>  [8,]  1.53423634
#>  [9,]  0.97866337
#> [10,] -1.71655895
#> [11,] -1.81529432
#> [12,] -0.96174076
#> [13,]  3.84147911
#> [14,]  3.73851015
#> [15,]  1.67298589
#> [16,]  3.27016257
#> [17,] -3.37647629
#> [18,] -0.11179006
#> [19,] -1.49183817
#> [20,] -0.13707747
#> [21,] -3.78182322
#> 
#> $fitted.values
#>           [,1]
#>  [1,] 21.90274
#>  [2,] 26.66002
#>  [3,] 21.42287
#>  [4,] 16.14828
#>  [5,] 22.55757
#>  [6,] 18.56059
#>  [7,] 19.69136
#>  [8,] 14.86576
#>  [9,] 16.32134
#> [10,] 16.91656
#> [11,] 12.21529
#> [12,] 11.36174
#> [13,] 10.85852
#> [14,] 28.66149
#> [15,] 28.72701
#> [16,] 30.62984
#> [17,] 24.87648
#> [18,] 15.61179
#> [19,] 16.69184
#> [20,] 15.13708
#> [21,] 25.18182
#> 
#> $lenb
#> [1] 19
#> 
#> $coefficients
#>           [,1]
#> [1,] 17.377316
#> [2,] -3.172498
#> [3,]  1.884618
#> [4,]  3.961995
#> 
#> $x
#>       [,1]  [,2] [,3] [,4]
#>  [1,]    1 1.260 2.42    1
#>  [2,]    1 0.705 4.01    1
#>  [3,]    1 1.600 4.84    0
#>  [4,]    1 1.825 2.42    0
#>  [5,]    1 1.575 5.40    0
#>  [6,]    1 1.825 3.70    0
#>  [7,]    1 1.825 4.30    0
#>  [8,]    1 2.455 2.80    0
#>  [9,]    1 2.115 3.00    0
#> [10,]    1 2.165 3.40    0
#> [11,]    1 3.635 3.38    0
#> [12,]    1 3.809 3.22    0
#> [13,]    1 3.730 2.82    0
#> [14,]    1 0.585 4.87    1
#> [15,]    1 0.000 3.92    1
#> [16,]    1 0.220 5.30    1
#> [17,]    1 0.850 5.41    0
#> [18,]    1 1.905 2.27    0
#> [19,]    1 1.820 2.70    0
#> [20,]    1 1.955 0.00    1
#> [21,]    1 1.165 4.00    1
#> 
#> attr(,"class")
#> [1] "mars"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.720634