Generalized Additive Quantile Regression Model for Multiple Quantiles

Quantile Regression with generalized additive models for fitting a learner on multiple quantiles simultaneously. Calls qgam::mqgam() from package qgam.

Form

For the form parameter, a gam formula specific to the Task is required (see example and ?mgcv::formula.gam). If no formula is provided, a fallback formula using all features in the task is used that will make the Learner behave like Linear Quantile Regression. The features specified in the formula need to be the same as columns with col_roles "feature" in the task.

Quantile

The quantiles for the Learner, i.e. qu parameter from qgam::mqgam(), is set using the values specified in learner$quantiles. The response quantile can be specified using learner$quantile_response.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.mqgam")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”, “quantiles”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, qgam

Parameters

Id	Type	Default	Levels	Range

form	untyped	-		-

lsig	numeric	-		\((-\infty, \infty)\)

err	numeric	-		\([0, 1]\)

cluster	untyped	NULL		-

multicore	logical	-	TRUE, FALSE	-

ncores	numeric	-		\((-\infty, \infty)\)

paropts	untyped	list()		-

link	untyped	"identity"		-

argGam	untyped	-		-

discrete	logical	FALSE	TRUE, FALSE	-

References

Fasiolo, Matteo, Wood, N. S, Zaffran, Margaux, Nedellec, Raphael, Goude, Yannig (2017). “Fast Calibrated Additive Quantile Regression.” Journal of the American Statistical Association, 116, 1402–1412. doi:10.1080/01621459.2020.1725521 .

See also

Author

lona-k

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrMQGam

Methods

Public methods

LearnerRegrMQGam$new()

LearnerRegrMQGam$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrMQGam$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrMQGam$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.