Regression Priority Lasso Learner

Patient outcome prediction based on multi-omics data taking practitioners’ preferences into account. Calls prioritylasso::prioritylasso() from prioritylasso.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.priority_lasso")
lrn("regr.priority_lasso")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, prioritylasso

Parameters

Id	Type	Default	Levels	Range

blocks	untyped	-		-

max.coef	untyped	NULL		-

block1.penalization	logical	TRUE	TRUE, FALSE	-

lambda.type	character	lambda.min	lambda.min, lambda.1se	-

standardize	logical	TRUE	TRUE, FALSE	-

nfolds	integer	5		\([1, \infty)\)

foldid	untyped	NULL		-

cvoffset	logical	FALSE	TRUE, FALSE	-

cvoffsetnfolds	integer	10		\([1, \infty)\)

handle.missingtestdata	character	-	none, omit.prediction, set.zero, impute.block	-

include.allintercepts	logical	FALSE	TRUE, FALSE	-

use.blocks	untyped	"all"		-

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

scale.y	logical	FALSE	TRUE, FALSE	-

return.x	logical	TRUE	TRUE, FALSE	-

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, 1]\)

References

Simon K, Vindi J, Roman H, Tobias H, Anne-Laure B (2018). “Priority-Lasso: a simple hierarchical approach to the prediction of clinical outcome using multi-omics data.” BMC Bioinformatics, 19. doi:10.1186/s12859-018-2344-6 .

See also

Author

HarutyunyanLiana

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrPriorityLasso

Methods

Public methods

LearnerRegrPriorityLasso$new()

LearnerRegrPriorityLasso$selected_features()

LearnerRegrPriorityLasso$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrPriorityLasso$new()

Method selected_features()

Selected features when coef is positive

Usage

LearnerRegrPriorityLasso$selected_features()

Returns

character().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrPriorityLasso$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("regr.priority_lasso")
print(learner)
#> <LearnerRegrPriorityLasso:regr.priority_lasso>: Priority Lasso
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, prioritylasso
#> * Predict Types:  [response]
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# available parameters:
learner$param_set$ids()
#>  [1] "blocks"                 "max.coef"               "block1.penalization"   
#>  [4] "lambda.type"            "standardize"            "nfolds"                
#>  [7] "foldid"                 "cvoffset"               "cvoffsetnfolds"        
#> [10] "handle.missingtestdata" "include.allintercepts"  "use.blocks"            
#> [13] "alignment"              "alpha"                  "big"                   
#> [16] "devmax"                 "dfmax"                  "eps"                   
#> [19] "epsnr"                  "exclude"                "exmx"                  
#> [22] "fdev"                   "gamma"                  "grouped"               
#> [25] "intercept"              "keep"                   "lambda"                
#> [28] "lambda.min.ratio"       "lower.limits"           "maxit"                 
#> [31] "mnlam"                  "mxit"                   "mxitnr"                
#> [34] "nlambda"                "offset"                 "parallel"              
#> [37] "penalty.factor"         "pmax"                   "pmin"                  
#> [40] "prec"                   "standardize.response"   "thresh"                
#> [43] "trace.it"               "type.gaussian"          "type.logistic"         
#> [46] "type.multinomial"       "upper.limits"           "scale.y"               
#> [49] "return.x"               "predict.gamma"          "relax"                 
#> [52] "s"