Accelerated Oblique Random Survival Forest Learner

Accelerated oblique random survival forest. Calls aorsf::orsf() from aorsf. Note that although the learner has the property "missing" and it can in principle deal with missing values, the behaviour has to be configured using the parameter na_action.

Initial parameter values

n_thread: This parameter is initialized to 1 (default is 0) to avoid conflicts with the mlr3 parallelization.

mtry:

This hyperparameter can alternatively be set via the added hyperparameter mtry_ratio as mtry = max(ceiling(mtry_ratio * n_features), 1). Note that mtry and mtry_ratio are mutually exclusive.

Prediction types

This learner returns three prediction types:

distr: a survival matrix in two dimensions, where observations are represented in rows and (unique event) time points in columns. Calculated using the internal predict.ObliqueForest() function.

response: the restricted mean survival time of each test observation, derived from the survival matrix prediction (distr).

crank: the expected mortality using mlr3proba::.surv_return().

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.aorsf")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “response”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, mlr3extralearners, aorsf

Parameters

Id	Type	Default	Levels	Range

n_tree	integer	500		\([1, \infty)\)

n_split	integer	5		\([1, \infty)\)

n_retry	integer	3		\([0, \infty)\)

n_thread	integer	-		\([0, \infty)\)

pred_aggregate	logical	TRUE	TRUE, FALSE	-

pred_simplify	logical	FALSE	TRUE, FALSE	-

oobag	logical	FALSE	TRUE, FALSE	-

mtry	integer	NULL		\([1, \infty)\)

mtry_ratio	numeric	-		\([0, 1]\)

sample_with_replacement	logical	TRUE	TRUE, FALSE	-

sample_fraction	numeric	0.632		\([0, 1]\)

control_type	character	fast	fast, cph, net, custom	-

control_custom_fun	untyped	-		-

split_rule	character	logrank	logrank, cstat	-

control_fast_do_scale	logical	FALSE	TRUE, FALSE	-

control_fast_ties	character	efron	efron, breslow	-

control_cph_ties	character	efron	efron, breslow	-

control_cph_eps	numeric	1e-09		\([0, \infty)\)

control_cph_iter_max	integer	20		\([1, \infty)\)

control_net_alpha	numeric	0.5		\((-\infty, \infty)\)

control_net_df_target	integer	NULL		\([1, \infty)\)

leaf_min_events	integer	1		\([1, \infty)\)

leaf_min_obs	integer	5		\([1, \infty)\)

split_min_events	integer	5		\([1, \infty)\)

split_min_obs	integer	10		\([1, \infty)\)

split_min_stat	numeric	NULL		\([0, \infty)\)

oobag_pred_type	character	risk	none, surv, risk, chf, mort	-

importance	character	anova	none, anova, negate, permute	-

importance_max_pvalue	numeric	0.01		\([1e-04, 0.9999]\)

tree_seeds	integer	NULL		\([1, \infty)\)

oobag_pred_horizon	numeric	NULL		\([0, \infty)\)

oobag_eval_every	integer	NULL		\([1, \infty)\)

oobag_fun	untyped	NULL		-

attach_data	logical	TRUE	TRUE, FALSE	-

verbose_progress	logical	FALSE	TRUE, FALSE	-

na_action	character	fail	fail, impute_meanmode	-

References

Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min Y, Mcclure LA, Howard G, Simon N (2019). “Oblique random survival forests.” The Annals of Applied Statistics, 13(3). doi:10.1214/19-aoas1261 .

Jaeger BC, Welden S, Lenoir K, Speiser JL, Segar MW, Pandey A, Pajewski NM (2023). “Accelerated and interpretable oblique random survival forests.” Journal of Computational and Graphical Statistics, 1–16. doi:10.1080/10618600.2023.2231048 .

See also

Author

bcjaeger

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvAorsf

Methods

Public methods

LearnerSurvAorsf$new()

LearnerSurvAorsf$oob_error()

LearnerSurvAorsf$importance()

LearnerSurvAorsf$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvAorsf$new()

Method oob_error()

OOB concordance error extracted from the model slot eval_oobag$stat_values

Usage

LearnerSurvAorsf$oob_error()

Returns

numeric().

Method importance()

The importance scores are extracted from the model.

Usage

LearnerSurvAorsf$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvAorsf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.aorsf")
print(learner)
#> 
#> ── <LearnerSurvAorsf> (surv.aorsf): Oblique Random Forest ──────────────────────
#> • Model: -
#> • Parameters: n_thread=1
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and aorsf
#> • Predict Types: [crank], distr, and response
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, oob_error, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> ---------- Oblique random survival forest
#> 
#>      Linear combinations: Accelerated Cox regression
#>           N observations: 670
#>                 N events: 220
#>                  N trees: 500
#>       N predictors total: 6
#>    N predictors per node: 3
#>  Average leaves per tree: 40.774
#> Min observations in leaf: 5
#>       Min events in leaf: 1
#>           OOB stat value: 0.83
#>            OOB stat type: Harrell's C-index
#>      Variable importance: anova
#> 
#> -----------------------------------------
print(learner$importance())
#> revascdays     revasc        los        age   stchange      sysbp 
#>  0.5902859  0.4960548  0.2211467  0.2021641  0.1188305  0.1117324 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8612961