BlockForest Survival Learner

Random survival forests for blocks of clinical and omics covariate data. Calls blockForest::blockfor() from package blockForest.

In this learner, only the trained forest object ($forest) is retained. The optimized block-specific tuning parameters (paramvalues) and the biased OOB error estimate (biased_oob_error_donotuse) are discarded, as they are either not needed for downstream use or not reliable for performance estimation.

Prediction types

This learner returns two prediction types:

distr: a survival matrix in two dimensions, where observations are represented in rows and (unique event) time points in columns. Calculated using the internal blockForest:::predict.blockForest() function.

crank: the expected mortality using mlr3proba::surv_return().

Initial parameter values

num.threads is initialized to 1 to avoid conflicts with parallelization via future.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.blockforest")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, mlr3extralearners, blockForest

Parameters

Id	Type	Default	Levels	Range

blocks	untyped	-		-

block.method	character	BlockForest	BlockForest, RandomBlock, BlockVarSel, VarProb, SplitWeights	-

num.trees	integer	2000		\([1, \infty)\)

mtry	untyped	NULL		-

nsets	integer	300		\([1, \infty)\)

num.trees.pre	integer	1500		\([1, \infty)\)

splitrule	character	extratrees	logrank, extratrees, C, maxstat	-

always.select.block	integer	0		\([0, 1]\)

importance	character	-	none, impurity, impurity_corrected, permutation	-

num.threads	integer	-		\([1, \infty)\)

seed	integer	NULL		\((-\infty, \infty)\)

verbose	logical	TRUE	TRUE, FALSE	-

References

Hornung, R., Wright, N. M (2019). “Block Forests: Random forests for blocks of clinical and omics covariate data.” BMC Bioinformatics, 20(1), 1–17. doi:10.1186/s12859-019-2942-y , https://doi.org/10.1186/s12859-019-2942-y.

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvBlockForest

Methods

Public methods

LearnerSurvBlockForest$new()

LearnerSurvBlockForest$importance()

LearnerSurvBlockForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvBlockForest$new()

Method importance()

The importance scores are extracted from the model slot variable.importance.

Usage

LearnerSurvBlockForest$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvBlockForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define a Task
task = tsk("grace")
# Create train and test set
ids = partition(task)
# check task's features
task$feature_names
#> [1] "age"        "los"        "revasc"     "revascdays" "stchange"  
#> [6] "sysbp"     
# partition features to 2 blocks
blocks = list(bl1 = 1:3, bl2 = 4:6)
# define learner
learner = lrn("surv.blockforest", blocks = blocks,
              importance = "permutation", nsets = 10,
              num.trees = 50, num.trees.pre = 10, splitrule = "logrank")
# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
# feature importance
learner$importance()
#>  revascdays      revasc         age         los       sysbp    stchange 
#> 0.113555203 0.067798587 0.040618497 0.025714966 0.005125523 0.004791427 
# Make predictions for the test observations
pred = learner$predict(task, row_ids = ids$test)
pred
#> 
#> ── <PredictionSurv> for 330 observations: ──────────────────────────────────────
#>  row_ids time status     crank     distr
#>        1  180  FALSE 14.771953 <list[1]>
#>        5  180  FALSE 12.516166 <list[1]>
#>        9  180  FALSE  7.680491 <list[1]>
#>      ---  ---    ---       ---       ---
#>      992   26   TRUE 97.653554 <list[1]>
#>      994    7  FALSE 86.265001 <list[1]>
#>      997    2   TRUE 97.830893 <list[1]>
# Score the predictions
pred$score()
#> surv.cindex 
#>   0.8051583