Survival Conditional Random Forest Learner

A random forest based on conditional inference trees (ctree). Calls partykit::cforest() from partykit.

Prediction types

This learner returns two prediction types:

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using the internal partykit::predict.cforest() function.

crank: the expected mortality using mlr3proba::.surv_return().

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cforest")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, mlr3extralearners, partykit, sandwich, coin

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

replace	logical	FALSE	TRUE, FALSE	-

fraction	numeric	0.632		\([0, 1]\)

mtry	integer	-		\([0, \infty)\)

mtryratio	numeric	-		\([0, 1]\)

applyfun	untyped	-		-

cores	integer	NULL		\((-\infty, \infty)\)

trace	logical	FALSE	TRUE, FALSE	-

cluster	untyped	-		-

na.action	untyped	"stats::na.pass"		-

scores	untyped	-		-

teststat	character	quadratic	quadratic, maximum	-

splitstat	character	quadratic	quadratic, maximum	-

splittest	logical	FALSE	TRUE, FALSE	-

testtype	character	Univariate	Bonferroni, MonteCarlo, Univariate, Teststatistic	-

nmax	untyped	-		-

alpha	numeric	0.05		\([0, 1]\)

mincriterion	numeric	0.95		\([0, 1]\)

logmincriterion	numeric	-0.05129329		\((-\infty, \infty)\)

minsplit	integer	20		\([1, \infty)\)

minbucket	integer	7		\([1, \infty)\)

minprob	numeric	0.01		\([0, 1]\)

stump	logical	FALSE	TRUE, FALSE	-

lookahead	logical	FALSE	TRUE, FALSE	-

MIA	logical	FALSE	TRUE, FALSE	-

nresample	integer	9999		\([1, \infty)\)

tol	numeric	1.490116e-08		\([0, \infty)\)

maxsurrogate	integer	0		\([0, \infty)\)

numsurrogate	logical	FALSE	TRUE, FALSE	-

maxdepth	integer	Inf		\([0, \infty)\)

multiway	logical	FALSE	TRUE, FALSE	-

splittry	integer	2		\([0, \infty)\)

intersplit	logical	FALSE	TRUE, FALSE	-

majority	logical	FALSE	TRUE, FALSE	-

caseweights	logical	TRUE	TRUE, FALSE	-

saveinfo	logical	FALSE	TRUE, FALSE	-

update	logical	FALSE	TRUE, FALSE	-

splitflavour	character	ctree	ctree, exhaustive	-

maxvar	integer	-		\([1, \infty)\)

OOB	logical	FALSE	TRUE, FALSE	-

simplify	logical	TRUE	TRUE, FALSE	-

scale	logical	TRUE	TRUE, FALSE	-

maxpts	integer	25000		\((-\infty, \infty)\)

abseps	numeric	0.001		\([0, \infty)\)

releps	numeric	0		\([0, \infty)\)

Initial parameter values

cores: This parameter is initialized to 1 (default is NULL) to avoid threading conflicts with future.

Custom mlr3 parameters

mtry:

This hyperparameter can alternatively be set via the added hyperparameter mtryratio as mtry = max(ceiling(mtryratio * n_features), 1). Note that mtry and mtryratio are mutually exclusive.

References

Hothorn T, Zeileis A (2015). “partykit: A Modular Toolkit for Recursive Partytioning in R.” Journal of Machine Learning Research, 16(118), 3905-3909. http://jmlr.org/papers/v16/hothorn15a.html.

Hothorn T, Hornik K, Zeileis A (2006). “Unbiased Recursive Partitioning: A Conditional Inference Framework.” Journal of Computational and Graphical Statistics, 15(3), 651–674. doi:10.1198/106186006x133933 , https://doi.org/10.1198/106186006x133933.

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCForest

Methods

Public methods

LearnerSurvCForest$new()

LearnerSurvCForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCForest$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("rats")
learner = lrn("surv.cforest", ntree = 50)
splits = partition(task)
learner$train(task, splits$train)
pred = learner$predict(task, splits$test)