Survival Conditional Inference Tree Learner

Calls partykit::ctree from package partykit.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("surv.ctree")
lrn("surv.ctree")

Meta Information

Task type: “surv”

Predict Types: “distr”, “crank”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3extralearners, partykit, coin, sandwich

Parameters

Id	Type	Default	Levels	Range

teststat	character	quadratic	quadratic, maximum	\((-\infty, \infty)\)

splitstat	character	quadratic	quadratic, maximum	\((-\infty, \infty)\)

splittest	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

testtype	character	Bonferroni	Bonferroni, MonteCarlo, Univariate, Teststatistic	\((-\infty, \infty)\)

nmax	list	-		\((-\infty, \infty)\)

alpha	numeric	0.05		\([0, 1]\)

mincriterion	numeric	0.95		\([0, 1]\)

logmincriterion	numeric	-		\((-\infty, \infty)\)

minsplit	integer	20		\([1, \infty)\)

minbucket	integer	7		\([1, \infty)\)

minprob	numeric	0.01		\([0, \infty)\)

stump	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

lookahead	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

MIA	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

nresample	integer	9999		\([1, \infty)\)

tol	numeric	-		\([0, \infty)\)

maxsurrogate	integer	0		\([0, \infty)\)

numsurrogate	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

mtry	integer	Inf		\([0, \infty)\)

maxdepth	integer	Inf		\([0, \infty)\)

maxvar	integer	-		\([1, \infty)\)

multiway	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

splittry	integer	2		\([0, \infty)\)

intersplit	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

majority	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

caseweights	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

applyfun	list	-		\((-\infty, \infty)\)

cores	integer	NULL		\((-\infty, \infty)\)

saveinfo	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

update	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

splitflavour	character	ctree	ctree, exhaustive	\((-\infty, \infty)\)

offset	list	-		\((-\infty, \infty)\)

cluster	list	-		\((-\infty, \infty)\)

scores	list	-		\((-\infty, \infty)\)

doFit	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

maxpts	integer	25000		\((-\infty, \infty)\)

abseps	numeric	0.001		\([0, \infty)\)

releps	numeric	0		\([0, \infty)\)

References

Hothorn T, Zeileis A (2015). “partykit: A Modular Toolkit for Recursive Partytioning in R.” Journal of Machine Learning Research, 16(118), 3905-3909. http://jmlr.org/papers/v16/hothorn15a.html

Hothorn T, Hornik K, Zeileis A (2006). “Unbiased Recursive Partitioning: A Conditional Inference Framework.” Journal of Computational and Graphical Statistics, 15(3), 651–674. doi: 10.1198/106186006x133933

See also

Author

adibender

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCTree

Methods

Public methods

LearnerSurvCTree$new()

LearnerSurvCTree$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCTree$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCTree$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("partykit", quietly = TRUE) && requireNamespace("coin", quietly = TRUE) && requireNamespace("sandwich", quietly = TRUE)) {
  learner = mlr3::lrn("surv.ctree")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerSurvCTree:surv.ctree>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3extralearners, partykit, coin, sandwich
#> * Predict Type: distr
#> * Feature types: integer, numeric, factor, ordered
#> * Properties: weights
#>  [1] "teststat"        "splitstat"       "splittest"       "testtype"       
#>  [5] "nmax"            "alpha"           "mincriterion"    "logmincriterion"
#>  [9] "minsplit"        "minbucket"       "minprob"         "stump"          
#> [13] "lookahead"       "MIA"             "nresample"       "tol"            
#> [17] "maxsurrogate"    "numsurrogate"    "mtry"            "maxdepth"       
#> [21] "maxvar"          "multiway"        "splittry"        "intersplit"     
#> [25] "majority"        "caseweights"     "applyfun"        "cores"          
#> [29] "saveinfo"        "update"          "splitflavour"    "offset"         
#> [33] "cluster"         "scores"          "doFit"           "maxpts"         
#> [37] "abseps"          "releps"