Survival DNNSurv Learner

Calls survivalmodels::dnnsurv from package survivalmodels.

Details

Custom nets can be used in this learner either using the survivalmodels::build_keras_net utility function or using keras. The number of output channels should be of length 1 and number of input channels is the number of features plus number of cuts.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("surv.dnnsurv")
lrn("surv.dnnsurv")

Traits

Packages: survivalmodels, keras, pseudo, tensorflow, distr6

Predict Types: crank, distr

Feature Types: integer, numeric

Properties:

Custom mlr3 defaults

verbose:

Actual default: 1L

Adjusted default: 0L

Reason for change: Prevents plotting.

References

Zhao, L., & Feng, D. (2020). DNNSurv: Deep Neural Networks for Survival Analysis Using Pseudo Values. https://arxiv.org/abs/1908.02337

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvDNNSurv

Methods

Public methods

LearnerSurvDNNSurv$new()

LearnerSurvDNNSurv$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvDNNSurv$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvDNNSurv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("surv.dnnsurv"))
print(learner)

#> <LearnerSurvDNNSurv:surv.dnnsurv>
#> * Model: -
#> * Parameters: verbose=0
#> * Packages: survivalmodels, keras, pseudo, tensorflow, distr6
#> * Predict Type: crank
#> * Feature types: integer, numeric
#> * Properties: -

# available parameters:
learner$param_set$ids()

#>  [1] "cuts"                 "cutpoints"            "custom_model"        
#>  [4] "optimizer"            "lr"                   "beta_1"              
#>  [7] "beta_2"               "epsilon"              "decay"               
#> [10] "clipnorm"             "clipvalue"            "schedule_decay"      
#> [13] "momentum"             "nesterov"             "loss_weights"        
#> [16] "weighted_metrics"     "early_stopping"       "min_delta"           
#> [19] "patience"             "verbose"              "baseline"            
#> [22] "restore_best_weights" "batch_size"           "epochs"              
#> [25] "validation_split"     "shuffle"              "sample_weight"       
#> [28] "initial_epoch"        "steps_per_epoch"      "validation_steps"    
#> [31] "steps"                "callbacks"