Survival Flexible Parametric Spline Learner

Flexible parametric spline learner. Calls flexsurv::flexsurvspline() from flexsurv.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.flexible")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3proba, mlr3extralearners, flexsurv

Parameters

Id	Type	Default	Levels	Range

formula	untyped	-		-

bhazard	untyped	-		-

k	integer	0		\([0, \infty)\)

knots	untyped	-		-

bknots	untyped	-		-

scale	character	hazard	hazard, odds, normal	-

timescale	character	log	log, identity	-

spline	character	rp	rp, splines2ns	-

rtrunc	untyped	-		-

inits	untyped	-		-

fixedpars	untyped	-		-

cl	numeric	0.95		\([0, 1]\)

anc	untyped	-		-

maxiter	integer	30		\((-\infty, \infty)\)

rel.tolerance	numeric	1e-09		\((-\infty, \infty)\)

toler.chol	numeric	1e-10		\((-\infty, \infty)\)

debug	integer	0		\([0, 1]\)

outer.max	integer	10		\((-\infty, \infty)\)

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using flexsurv::flexsurvspline() and the estimated coefficients. For fitted coefficients, \(\hat{\beta} = (\hat{\beta_0},...,\hat{\beta_P})\), and the test data covariates \(X^T = (X_0,...,X_P)^T\), where \(X_0\) is a column of \(1\)s and \(\hat{\beta_0} = \hat{\gamma_0}\), the linear predictor vector is \(lp = \hat{\beta} X^T\).

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using predict.flexsurvreg(type = "survival", ...).

crank: same as lp.

response: mean survival time calculated using predict.flexsurvreg(type = "response", ...)

Initial parameter values

k:

Actual default: 0

Initial value: 1

Reason for change: The default value of 0 is equivalent to, and a much less efficient implementation of, LearnerSurvParametric.

References

Royston, Patrick, Parmar, KB M (2002). “Flexible parametric proportional-hazards and proportional-odds models for censored survival data, with application to prognostic modelling and estimation of treatment effects.” Statistics in medicine, 21(15), 2175–2197.

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvFlexible

Methods

Public methods

LearnerSurvFlexible$new()

LearnerSurvFlexible$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvFlexible$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvFlexible$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(survival)

# Define the task, split to train/test set
task = tsk("lung")
part = partition(task)

# Define the learner
learner = lrn("surv.flexible", k = 1,
  formula = Surv(time, status) ~ age + ph.karno + sex,
  anc = list(gamma1 = ~ sex))

# Train the learner on the training ids
learner$train(task, part$train)
#> Error in optim(method = "BFGS", par = c(gamma0 = 0.769559459031002, gamma1 = 0.809471579752982, gamma2 = -0.0867507621158821, age = 0.0112449832480452, ph.karno = -0.0412933200469301, sexm = 0.632805840429337, `gamma1(sexm)` = 0.04316080335661),     fn = function (optpars, ...)     {        pars[insert.locations] <- optpars        raw.pars <- pars        pars <- as.list(pars)        pars.event <- pars.nevent <- pars        if (npars > nbpars) {            beta <- raw.pars[(nbpars + 1):npars]            for (i in dlist$pars) {                pars[[i]] <- pars[[i]] + X[, mx[[i]], drop = FALSE] %*%                   beta[mx[[i]]]                pars.event[[i]] <- pars[[i]][event]                pars.nevent[[i]] <- pars[[i]][!event]            }        }        fnargs <- c(par.transform(pars), aux.pars)        fnargs.event <- c(par.transform(pars.event), aux.pars)        fnargs.nevent <- c(par.transform(pars.nevent), aux.pars)        dargs <- fnargs.event        dargs$x <- event.times        dargs$log <- TRUE        logdens <- call_distfn_quiet(dfns$d, dargs)        if (!all(event)) {            pmaxargs <- fnargs.nevent            pmaxargs$q <- lcens.times            pmax <- call_distfn_quiet(dfns$p, pmaxargs)            pmax[pmaxargs$q == Inf] <- 1            pargs <- fnargs.nevent            pargs$q <- rcens.times            pmin <- call_distfn_quiet(dfns$p, pargs)        }        targs <- fnargs        targs$q <- Y[, "start"]        plower <- call_distfn_quiet(dfns$p, targs)        targs$q <- rtrunc        pupper <- call_distfn_quiet(dfns$p, targs)        pupper[rtrunc == Inf] <- 1        pobs <- pupper - plower        if (do.bhazard) {            pargs <- fnargs.event            pargs$q <- event.times            pminb <- call_distfn_quiet(dfns$p, pargs)            logsurv_excess <- log(1 - pminb)            loghaz_excess <- logdens - logsurv_excess            haz_excess <- exp(loghaz_excess)            logdens_offset <- log(1 + bhazard[event]/haz_excess)            if (!all(event)) {                b_condsurv <- 1 - bhazard[!event]                b_condsurv[lcens.times == Inf] <- 0            }            if (any(is.finite(rtrunc)))                 stop("models with both right truncation and background hazards not supported")        }        else {            logdens_offset <- 0        }        loglik[event] <- (logdens + logdens_offset)        if (!all(event)) {            if (do.bhazard)                 loglik[!event] <- log((pmax - 1) * b_condsurv +                   1 - pmin)            else loglik[!event] <- log(pmax - pmin)        }        loglik <- loglik - log(pobs)        ret <- -sum(loglik * weights)        attr(ret, "indiv") <- loglik        ret    }, gr = function (optpars, Y, X = 0, weights, bhazard, rtrunc,         dlist, inits, dfns, aux, mx, fixedpars = NULL)     {        pars <- inits        npars <- length(pars)        pars[setdiff(seq_len(npars), fixedpars)] <- optpars        nbpars <- length(dlist$pars)        pars <- as.list(pars)        ncovs <- length(pars) - length(dlist$pars)        if (ncovs > 0)             beta <- unlist(pars[(nbpars + 1):npars])        for (i in dlist$pars) {            if (length(mx[[i]]) > 0)                 pars[[i]] <- pars[[i]] + X[, mx[[i]], drop = FALSE] %*%                   beta[mx[[i]]]            else pars[[i]] <- rep(pars[[i]], length(Y[, "stop"]))        }        dead <- Y[, "status"] == 1        ddcall <- list(t = Y[dead, "stop"])        dsccall <- list(t = Y[!dead, "stop"])        dstcall <- list(t = Y[, "start"])        for (i in 1:nbpars) ddcall[[names(pars)[i]]] <- dsccall[[names(pars)[i]]] <- dstcall[[names(pars)[i]]] <- dlist$inv.transforms[[i]](pars[[i]])        for (i in seq_along(aux)) {            ddcall[[names(aux)[i]]] <- dsccall[[names(aux)[i]]] <- dstcall[[names(aux)[i]]] <- aux[[i]]        }        for (i in dlist$pars) {            ddcall[[i]] <- ddcall[[i]][dead]            dsccall[[i]] <- dsccall[[i]][!dead]        }        nobs <- nrow(Y)        dloglik <- matrix(nrow = nobs, ncol = npars)        dloglik[dead, ] <- dderiv(dfns$DLd, ddcall, X[dead, ,             drop = FALSE], mx, dlist)        dloglik[!dead, ] <- dderiv(dfns$DLS, dsccall, X[!dead,             , drop = FALSE], mx, dlist)        dstrunc <- dderiv(dfns$DLS, dstcall, X, mx, dlist)        dloglik <- dloglik - dstrunc        if (any(bhazard > 0)) {            dcall <- ddcall            dcall$x <- ddcall$t            dcall$t <- NULL            dens <- do.call(dfns$d, dcall)            pcall <- dcall            pcall$q <- pcall$x            pcall$x <- NULL            surv <- 1 - do.call(dfns$p, pcall)            haz <- dens/surv            offseti <- 1/(1 + bhazard[dead]/haz)            dscense <- dderiv(dfns$DLS, ddcall, X[dead, , drop = FALSE],                 mx, dlist)            doff <- -offseti * bhazard[dead] * (dloglik[dead,                 ] - dscense)/haz            dloglik[dead, ] <- dloglik[dead, ] + doff        }        res <- -colSums(dloglik * weights)        res[setdiff(1:npars, fixedpars)]    }, Y = structure(c(455, 210, 310, 361, 567, 613, 61, 301,     371, 118, 390, 473, 26, 107, 93, 731, 460, 153, 95, 303,     519, 643, 53, 246, 689, 5, 687, 345, 444, 60, 163, 65, 230,     840, 305, 705, 363, 176, 95, 167, 806, 284, 147, 163, 655,     245, 477, 450, 156, 529, 351, 15, 181, 283, 212, 288, 199,     550, 92, 60, 551, 353, 267, 511, 457, 201, 404, 222, 62,     458, 353, 31, 229, 156, 329, 291, 179, 376, 384, 268, 142,     413, 181, 285, 301, 348, 197, 382, 180, 269, 300, 284, 292,     332, 285, 110, 270, 225, 225, 276, 135, 59, 202, 235, 239,     221, 185, 222, 175, 197, 203, 174, 177, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1,     1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1,     0, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0,     1, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1,     1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 455, 210, 310, 361, 567,     613, 61, 301, 371, 118, 390, 473, 26, 107, 93, 731, 460,     153, 95, 303, 519, 643, 53, 246, 689, 5, 687, 345, 444, 60,     163, 65, 230, 840, 305, 705, 363, 176, 95, 167, 806, 284,     147, 163, 655, 245, 477, 450, 156, 529, 351, 15, 181, 283,     212, 288, 199, 550, 92, 60, 551, 353, 267, 511, 457, 201,     404, 222, 62, 458, 353, 31, 229, 156, 329, 291, 179, 376,     384, 268, 142, 413, 181, 285, 301, 348, 197, 382, 180, 269,     300, 284, 292, 332, 285, 110, 270, 225, 225, 276, 135, 59,     202, 235, 239, 221, 185, 222, 175, 197, 203, 174, 177, 455,     210, 310, 361, 567, 613, 61, 301, 371, 118, 390, 473, 26,     107, 93, 731, 460, 153, 95, 303, 519, 643, 53, 246, 689,     5, 687, 345, 444, 60, 163, 65, 230, 840, 305, 705, 363, 176,     95, 167, 806, 284, 147, 163, 655, 245, 477, 450, 156, 529,     351, 15, 181, 283, 212, 288, 199, 550, 92, 60, 551, 353,     267, 511, 457, 201, 404, 222, 62, 458, 353, 31, 229, 156,     329, 291, 179, 376, 384, 268, 142, 413, 181, 285, 301, 348,     197, 382, 180, 269, 300, 284, 292, 332, 285, 110, 270, 225,     225, 276, 135, 59, 202, 235, 239, 221, 185, 222, 175, 197,     203, 174, 177, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf), dim = c(113L, 6L    ), dimnames = list(c("1", "2", "3", "4", "5", "6", "7", "8",     "9", "10", "11", "12", "13", "14", "15", "16", "17", "18",     "19", "20", "21", "22", "23", "24", "25", "26", "27", "28",     "29", "30", "31", "32", "33", "34", "35", "36", "37", "38",     "39", "40", "41", "42", "43", "44", "45", "46", "47", "48",     "49", "50", "51", "52", "53", "54", "55", "56", "57", "58",     "59", "60", "61", "62", "63", "64", "65", "66", "67", "68",     "69", "70", "71", "72", "73", "74", "75", "76", "77", "78",     "79", "80", "81", "82", "83", "84", "85", "86", "87", "88",     "89", "90", "91", "92", "93", "94", "95", "96", "97", "98",     "99", "100", "101", "102", "103", "104", "105", "106", "107",     "108", "109", "110", "111", "112", "113"), c("time", "status",     "start", "stop", "time1", "time2")), type = "right"), X = structure(c(68,     57, 68, 71, 57, 70, 56, 67, 58, 70, 53, 69, 73, 60, 74, 64,     70, 73, 76, 74, 63, 74, 68, 58, 59, 65, 58, 64, 75, 65, 69,     68, 67, 63, 48, 51, 56, 73, 55, 44, 44, 71, 61, 72, 63, 57,     64, 69, 66, 54, 75, 69, 44, 80, 49, 66, 60, 69, 50, 64, 77,     47, 67, 74, 54, 73, 74, 76, 65, 57, 71, 82, 70, 55, 69, 62,     63, 56, 62, 44, 63, 64, 61, 65, 61, 58, 56, 43, 56, 74, 60,     39, 51, 45, 72, 64, 72, 64, 70, 52, 60, 73, 50, 63, 50, 67,     69, 65, 57, 67, 71, 66, 58, 90, 90, 70, 60, 80, 90, 60, 80,     90, 60, 80, 90, 60, 50, 50, 80, 80, 60, 60, 90, 80, 90, 90,     100, 90, 100, 80, 90, 70, 90, 80, 70, 80, 90, 80, 100, 80,     90, 70, 80, 80, 80, 100, 70, 100, 80, 90, 80, 80, 80, 60,     90, 80, 80, 70, 70, 70, 70, 80, 80, 80, 100, 90, 60, 90,     70, 80, 70, 80, 80, 100, 100, 70, 70, 70, 70, 80, 80, 90,     90, 90, 80, 90, 100, 90, 90, 90, 100, 60, 100, 100, 100,     90, 90, 70, 80, 80, 90, 100, 100, 90, 60, 100, 100, 60, 80,     90, 90, 80, 80, 80, 90, 80, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1,     1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0,     0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0,     1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1,     1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0,     1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0,     1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1,     0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0,     1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1,     0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1,     0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1,     0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1,     1, 1, 0, 1, 0, 1, 0), dim = c(113L, 4L), dimnames = list(        c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10",         "11", "12", "13", "14", "15", "16", "17", "18", "19",         "20", "21", "22", "23", "24", "25", "26", "27", "28",         "29", "30", "31", "32", "33", "34", "35", "36", "37",         "38", "39", "40", "41", "42", "43", "44", "45", "46",         "47", "48", "49", "50", "51", "52", "53", "54", "55",         "56", "57", "58", "59", "60", "61", "62", "63", "64",         "65", "66", "67", "68", "69", "70", "71", "72", "73",         "74", "75", "76", "77", "78", "79", "80", "81", "82",         "83", "84", "85", "86", "87", "88", "89", "90", "91",         "92", "93", "94", "95", "96", "97", "98", "99", "100",         "101", "102", "103", "104", "105", "106", "107", "108",         "109", "110", "111", "112", "113"), c("age", "ph.karno",         "sexm", gamma1 = "gamma1(sexm)"))), weights = c(1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,     1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1), bhazard = c(0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), rtrunc = c(Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf, Inf,     Inf, Inf, Inf, Inf), dlist = list(name = "survspline", pars = c("gamma0",     "gamma1", "gamma2"), location = "gamma0", transforms = list(        function (x)         x, function (x)         x, function (x)         x), inv.transforms = list(function (x)     x, function (x)     x, function (x)     x), inits = function (t, mf, mml, aux)     {        aux$skip.pd.check <- TRUE        inits <- flexsurv.splineinits(t = t, mf = mf, mml = mml,             aux = aux)        Y <- check.flexsurv.response(model.extract(mf, "response"))        intcens <- Y[, "status"] == 3 & (Y[, "start"] < Y[, "stop"])        Y[intcens, "status"] <- 1        Y[intcens, "stop"] <- Y[intcens, "start"] + (Y[intcens,             "stop"] - Y[intcens, "start"])/2        Y[, "status"] <- ifelse(Y[, "status"] == 1, 1, 0)        inc <- Y[, "start"] < Y[, "stop"]        X <- mml[[1]][, -1, drop = FALSE]        formdat <- as.data.frame(cbind(Y, X, weights = model.extract(mf,             "weights")))        names(formdat)[1:ncol(Y)] <- colnames(Y)        form <- c("Surv(start, stop, status) ~")        if (ncol(X) > 0) {            names(formdat)[ncol(Y) + 1:ncol(X)] <- paste0("X",                 1:ncol(X))            form <- paste(form, paste(paste0("X", 1:ncol(X)),                 collapse = " + "))            formdat <- formdat[inc, ]            cox <- coxph(as.formula(form), weights = weights,                 data = formdat)            covinds <- length(aux$knots) + 1:ncol(X)            inits[covinds] <- coef(cox)        }        inits    }), inits = c(gamma0 = 0.769559459031002, gamma1 = 0.809471579752982,     gamma2 = -0.0867507621158821, age = 0.0112449832480452, ph.karno = -0.0412933200469301,     sexm = 0.632805840429337, `gamma1(sexm)` = 0.04316080335661    ), dfns = list(p = function (q, beta = 0, X = 0, knots = c(-10,         10), scale = "hazard", timescale = "log", spline = "rp",         offset = 0, lower.tail = TRUE, log.p = FALSE, gamma0,         gamma1, gamma2)     {        base.fn <- function(q, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0, lower.tail = TRUE, log.p = FALSE) {            betax_warn(beta, X, offset)            d <- dbase.survspline(q = q, gamma = gamma, knots = knots,                 scale = scale, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            if (any(ind)) {                ret[ind][q == 0] <- 0                ret[ind][q == Inf] <- 1                finite <- q > 0 & q < Inf                ind[ind][!finite] <- FALSE                q <- q[finite]                gamma <- gamma[finite, , drop = FALSE]                knots <- knots[finite, , drop = FALSE]            }            if (any(ind)) {                eta <- rowSums(basis(knots, tsfn(q, timescale),                   spline = spline) * gamma) + as.numeric(X %*%                   beta)                surv <- Slink(scale)(eta)                ret[ind] <- as.numeric(1 - surv)            }            if (!lower.tail)                 ret <- 1 - ret            if (log.p)                 ret <- log(ret)            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(q = q, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, lower.tail = lower.tail, log.p = log.p,             gamma = gamma))    }, d = function (x, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         log = FALSE, gamma0, gamma1, gamma2)     {        base.fn <- function(x, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0, log = FALSE) {            betax_warn(beta, X, offset)            d <- dbase.survspline(q = x, gamma = gamma, knots = knots,                 scale = scale, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            if (any(ind)) {                eta <- rowSums(basis(knots, tsfn(q, timescale),                   spline = spline) * gamma)                eeta <- exp(ldlink(scale)(eta))                ret[ind][eeta == 0] <- 0                ret[ind][is.nan(eeta)] <- NaN                ind2 <- !(eeta == 0 | is.nan(eeta))                q <- q[ind2]                gamma <- gamma[ind2, , drop = FALSE]                knots <- knots[ind2, , drop = FALSE]                eeta <- eeta[ind2]                ind[ind] <- ind[ind] & ind2                dsum <- rowSums(dbasis(knots, tsfn(q, timescale),                   spline = spline) * gamma)                ret[ind] <- dtsfn(q, timescale) * dsum * eeta                ret[ind][ret[ind] <= 0] <- 0            }            if (log) {                ret <- log(ret)            }            as.numeric(ret)        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(x = x, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, log = log, gamma = gamma))    }, h = function (x, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         gamma0, gamma1, gamma2)     {        base.fn <- function(x, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0) {            betax_warn(beta, X, offset)            match.arg(scale, c("hazard", "odds", "normal"))            d <- dbase.survspline(q = x, gamma = gamma, knots = knots,                 scale = scale, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            if (any(ind)) {                eta <- rowSums(basis(knots, tsfn(q, timescale),                   spline = spline) * gamma) + as.numeric(X %*%                   beta)                eeta <- hlink(scale)(eta)                ret[ind] <- dtsfn(q, timescale) * rowSums(dbasis(knots,                   tsfn(q, timescale), spline = spline) * gamma) *                   eeta                ret[ind][ret[ind] <= 0] <- 0            }            as.numeric(ret)        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(x = x, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, gamma = gamma))    }, H = function (x, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         gamma0, gamma1, gamma2)     {        base.fn <- function(x, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0) {            match.arg(scale, c("hazard", "odds", "normal"))            d <- dbase.survspline(q = x, gamma = gamma, knots = knots,                 scale = scale, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            if (any(ind)) {                eta <- rowSums(basis(knots, tsfn(q, timescale),                   spline = spline) * gamma) + as.numeric(X %*%                   beta)                ret[ind] <- as.numeric(Hlink(scale)(eta))            }            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(x = x, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, gamma = gamma))    }, r = function (n, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         gamma0, gamma1, gamma2)     {        base.fn <- function(n, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0) {            betax_warn(beta, X, offset)            if (length(n) > 1)                 n <- length(n)            ret <- qsurvspline(p = runif(n), gamma = gamma, knots = knots,                 scale = scale, timescale = timescale, spline = spline)            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(n = n, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, gamma = gamma))    }, DLd = function (t, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", gamma0,         gamma1, gamma2)     {        base.fn <- function(t, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp") {            d <- dbase.survspline(q = t, gamma = gamma, knots = knots,                 scale = scale, deriv = 1, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            t <- q            b <- basis(knots, tsfn(t, timescale), spline = spline)            db <- dbasis(knots, tsfn(t, timescale), spline = spline)            eta <- rowSums(b * gamma) + as.numeric(X %*% beta)            ds <- rowSums(db * gamma)            npars <- ncol(gamma)            parnames <- paste0("gamma", seq_len(npars) - 1)            colnames(ret) <- parnames            for (i in 1:ncol(gamma)) {                if (scale == "hazard")                   ret[ind, i] <- db[, i]/ds + b[, i] * (1 - exp(eta))                else if (scale == "odds") {                  eeta <- 1 - 2 * exp(eta)/(1 + exp(eta))                  ret[ind, i] <- db[, i]/ds + b[, i] * eeta                }            }            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(t = t, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             gamma = gamma))    }, DLS = function (t, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", gamma0,         gamma1, gamma2)     {        base.fn <- function(t, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp") {            d <- dbase.survspline(q = t, gamma = gamma, knots = knots,                 scale = scale, deriv = 1, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            t <- q            b <- basis(knots, tsfn(t, timescale), spline = spline)            if (any(ind)) {                eta <- rowSums(b * gamma) + as.numeric(X %*%                   beta)                for (i in 1:ncol(gamma)) {                  if (scale == "hazard")                     ret[ind, i] <- ifelse(t == 0, 0, -b[, i] *                       exp(eta))                  else if (scale == "odds") {                    eeta <- exp(eta)/(1 + exp(eta))                    ret[ind, i] <- ifelse(t == 0, 0, -b[, i] *                       eeta)                  }                }            }            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(t = t, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             gamma = gamma))    }, D2Ld = function (t, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", gamma0,         gamma1, gamma2)     {        base.fn <- function(t, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp") {            d <- dbase.survspline(q = t, gamma = gamma, knots = knots,                 scale = scale, deriv = 2, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            t <- q            b <- basis(knots, tsfn(t, timescale), spline = spline)            db <- dbasis(knots, tsfn(t, timescale), spline = spline)            eta <- rowSums(b * gamma) + as.numeric(X %*% beta)            ds <- rowSums(db * gamma)            if (scale == "odds") {                eeta <- 2 * exp(eta)/(1 + exp(eta))^2                db <- dbasis(knots, tsfn(t, timescale), spline = spline)            }            npars <- ncol(gamma)            for (i in 1:npars) {                for (j in 1:npars) {                  if (scale == "hazard")                     ret[ind, i, j] <- -db[, i] * db[, j]/ds^2 -                       b[, i] * b[, j] * exp(eta)                  else if (scale == "odds")                     ret[ind, i, j] <- -db[, i] * db[, j]/ds^2 -                       b[, i] * b[, j] * eeta                }            }            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(t = t, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             gamma = gamma))    }, D2LS = function (t, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", gamma0,         gamma1, gamma2)     {        base.fn <- function(t, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp") {            tmp_gamma <- gamma            tmp_t <- t            tmp_knots <- knots            d <- dbase.survspline(q = t, gamma = gamma, knots = knots,                 scale = scale, deriv = 2, spline = spline)            for (i in seq_along(d)) assign(names(d)[i], d[[i]])            t <- q            b <- basis(knots, tsfn(t, timescale), spline = spline)            npars <- ncol(gamma)            if (length(t) > 0) {                eta <- rowSums(b * gamma) + as.numeric(X %*%                   beta)                if (scale == "odds") {                  eeta <- exp(eta)/(1 + exp(eta))^2                }                for (i in 1:npars) {                  for (j in 1:npars) {                    if (scale == "hazard")                       ret[ind, i, j] <- ifelse(t == 0, 0, -b[,                         i] * b[, j] * exp(eta))                    else if (scale == "odds")                       ret[ind, i, j] <- ifelse(t == 0, 0, -b[,                         i] * b[, j] * eeta)                  }                }            }            ret        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(t = t, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             gamma = gamma))    }, rmst = function (t, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         start = 0, gamma0, gamma1, gamma2)     {        base.fn <- function(t, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0, start = 0) {            betax_warn(beta, X, offset)            rmst_generic(psurvspline, t, start = start, matargs = c("gamma",                 "knots"), scalarargs = c("scale", "timescale",                 "spline"), gamma = gamma, knots = knots, scale = scale,                 timescale = timescale, spline = spline)        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(t = t, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, start = start, gamma = gamma))    }, mean = function (beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         gamma0, gamma1, gamma2)     {        base.fn <- function(gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0) {            betax_warn(beta, X, offset)            nt <- if (is.matrix(gamma))                 nrow(gamma)            else 1            rmst_generic(psurvspline, rep(Inf, nt), start = 0,                 matargs = c("gamma", "knots"), scalarargs = c("scale",                   "timescale", "spline"), gamma = gamma, knots = knots,                 scale = scale, timescale = timescale, spline = spline)        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, gamma = gamma))    }, q = function (p, beta = 0, X = 0, knots = c(-10, 10),         scale = "hazard", timescale = "log", spline = "rp", offset = 0,         lower.tail = TRUE, log.p = FALSE, gamma0, gamma1, gamma2)     {        base.fn <- function(p, gamma, beta = 0, X = 0, knots = c(-10,             10), scale = "hazard", timescale = "log", spline = "rp",             offset = 0, lower.tail = TRUE, log.p = FALSE) {            betax_warn(beta, X, offset)            if (log.p)                 p <- exp(p)            if (!lower.tail)                 p <- 1 - p            qgeneric(psurvspline, p = p, matargs = c("gamma",                 "knots"), scalarargs = c("scale", "timescale",                 "spline"), gamma = gamma, knots = knots, scale = scale,                 timescale = timescale, spline = spline)        }        gamma <- do.call("cbind", list(gamma0, gamma1, gamma2))        do.call(base.fn, list(p = p, beta = beta, X = X, knots = knots,             scale = scale, timescale = timescale, spline = spline,             offset = offset, lower.tail = lower.tail, log.p = log.p,             gamma = gamma))    }, deriv = TRUE, hessian = TRUE), aux = list(knots = c(1.6094379124341,     `50%` = 5.50125821054473, 6.59441345974978), scale = "hazard",         timescale = "log", spline = "rp"), mx = list(gamma0 = 1:3,         gamma1 = 4L, gamma2 = NULL), fixedpars = NULL, hessian = FALSE): initial value in 'vmmin' is not finite
print(learner$model)
#> NULL

# Make predictions for the test rows
predictions = learner$predict(task, part$test)
#> Error: Cannot predict, Learner 'surv.flexible' has not been trained yet
print(predictions)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'predictions' not found

# Score the predictions
predictions$score()
#> Error: object 'predictions' not found