Survival Gradient Boosting Machine Learner

Gradient Boosting for Survival Analysis. Calls gbm::gbm() from gbm.

Prediction types

This learner returns two prediction types, using the internal predict.gbm() function:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation.

crank: same as lp.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.gbm")

Meta Information

Task type: “surv”

Predict Types: “crank”, “lp”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, mlr3extralearners, gbm

Parameters

Id	Type	Default	Levels	Range

distribution	character	coxph	coxph	-

n.trees	integer	100		\([1, \infty)\)

cv.folds	integer	0		\([0, \infty)\)

interaction.depth	integer	1		\([1, \infty)\)

n.minobsinnode	integer	10		\([1, \infty)\)

shrinkage	numeric	0.001		\([0, \infty)\)

bag.fraction	numeric	0.5		\([0, 1]\)

train.fraction	numeric	1		\([0, 1]\)

keep.data	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

var.monotone	untyped	-		-

n.cores	integer	1		\((-\infty, \infty)\)

single.tree	logical	FALSE	TRUE, FALSE	-

Initial parameter values

distribution:

Actual default: "bernoulli"

Adjusted default: "coxph"

Reason for change: This is the only distribution available for survival.

keep.data:

Actual default: TRUE

Adjusted default: FALSE

Reason for change: keep.data = FALSE saves memory during model fitting.

n.cores:

Actual default: NULL

Adjusted default: 1

Reason for change: Suppressing the automatic internal parallelization if cv.folds > 0 and avoid threading conflicts with future.

References

Friedman, H J (2002). “Stochastic gradient boosting.” Computational statistics & data analysis, 38(4), 367–378.

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGBM

Methods

Public methods

LearnerSurvGBM$new()

LearnerSurvGBM$importance()

LearnerSurvGBM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGBM$new()

Method importance()

The importance scores are extracted from the model slot variable.importance.

Usage

LearnerSurvGBM$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.gbm")
print(learner)
#> 
#> ── <LearnerSurvGBM> (surv.gbm): Gradient Boosting ──────────────────────────────
#> • Model: -
#> • Parameters: distribution=coxph, keep.data=FALSE, n.cores=1
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and gbm
#> • Predict Types: [crank] and lp
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> gbm::gbm(formula = f, distribution = "coxph", data = task$data(), 
#>     weights = NULL, keep.data = FALSE, n.cores = 1L)
#> A gradient boosted model with coxph loss function.
#> 100 iterations were performed.
#> There were 6 predictors of which 6 had non-zero influence.
print(learner$importance())
#> revascdays     revasc        age      sysbp        los   stchange 
#>  43.322220  26.360843  15.139899   7.701870   7.107836   0.367332 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Using 100 trees...

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8688895