GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

newoffset	untyped	-		-

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.193900
#> 2   2  0.85 0.176700
#> 3   2  1.84 0.161000
#> 4   2  2.70 0.146700
#> 5   3  4.27 0.133600
#> 6   3  7.10 0.121800
#> 7   3  9.21 0.110900
#> 8   3 10.84 0.101100
#> 9   3 12.15 0.092110
#> 10  3 13.22 0.083920
#> 11  3 14.10 0.076470
#> 12  3 14.84 0.069680
#> 13  3 15.45 0.063490
#> 14  3 15.97 0.057850
#> 15  3 16.41 0.052710
#> 16  4 16.79 0.048020
#> 17  4 17.12 0.043760
#> 18  6 17.43 0.039870
#> 19  6 17.70 0.036330
#> 20  6 17.93 0.033100
#> 21  6 18.13 0.030160
#> 22  6 18.29 0.027480
#> 23  6 18.44 0.025040
#> 24  6 18.55 0.022820
#> 25  6 18.66 0.020790
#> 26  6 18.74 0.018940
#> 27  6 18.81 0.017260
#> 28  6 18.88 0.015730
#> 29  6 18.93 0.014330
#> 30  6 18.97 0.013060
#> 31  6 19.01 0.011900
#> 32  6 19.04 0.010840
#> 33  6 19.07 0.009876
#> 34  6 19.09 0.008999
#> 35  6 19.10 0.008199
#> 36  6 19.12 0.007471
#> 37  6 19.13 0.006807
#> 38  6 19.14 0.006203
#> 39  6 19.15 0.005652
#> 40  6 19.16 0.005150
#> 41  6 19.17 0.004692
#> 42  6 19.17 0.004275
#> 43  6 19.18 0.003895
#> 44  6 19.18 0.003549
#> 45  6 19.18 0.003234
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35   5      1          2        0   172
#>   [5,]  35  10      1          9        0   106
#>   [6,]  34   5      0          5        0   120
#>   [7,]  35   2      0        180        0   121
#>   [8,]  35   2      1          1        1   112
#>   [9,]  37   9      0        180        1   151
#>  [10,]  38  13      1          0        1   161
#>  [11,]  38   2      0        115        0   150
#>  [12,]  36   1      0        180        1   155
#>  [13,]  36   5      1          0        1   115
#>  [14,]  33   6      1          1        1   115
#>  [15,]  38  16      1         10        0   160
#>  [16,]  40  12      1          9        0   153
#>  [17,]  42   3      1          1        1   130
#>  [18,]  37   1      1          0        1   146
#>  [19,]  40   2      1          1        1   148
#>  [20,]  42   2      0        180        1   100
#>  [21,]  42   2      0          2        0   140
#>  [22,]  40   6      0        180        1   138
#>  [23,]  42   2      0        180        0   100
#>  [24,]  43   3      1          0        1   100
#>  [25,]  41   2      1          1        0   166
#>  [26,]  43   4      1          0        1   130
#>  [27,]  42   4      0        180        0   162
#>  [28,]  40   3      1          1        0   170
#>  [29,]  42   2      0        180        1   124
#>  [30,]  45   3      0        180        1   154
#>  [31,]  41  10      1          8        0   150
#>  [32,]  45   9      1          7        0   110
#>  [33,]  44   2      1          1        1   150
#>  [34,]  45   2      0        180        1   140
#>  [35,]  46  15      0        180        0   120
#>  [36,]  48  15      0        180        1   160
#>  [37,]  44   3      1          0        1   180
#>  [38,]  46   7      1          2        0   166
#>  [39,]  43  29      0        180        1   180
#>  [40,]  45   4      1          0        0   124
#>  [41,]  43  10      0        180        0   185
#>  [42,]  46  13      1         10        0   100
#>  [43,]  47   4      1          3        1   160
#>  [44,]  45   8      1          0        1   117
#>  [45,]  49   5      0         73        1   136
#>  [46,]  45   5      0          5        0   141
#>  [47,]  46   2      1          1        1   122
#>  [48,]  46   6      1          0        1   100
#>  [49,]  44   4      1          0        1   114
#>  [50,]  47   2      0        180        0   108
#>  [51,]  44   9      1          8        1   135
#>  [52,]  45   5      0        180        1   190
#>  [53,]  46   5      1          3        0   130
#>  [54,]  46   4      0        180        1   121
#>  [55,]  47   3      1          1        1   120
#>  [56,]  48  12      1         11        0   200
#>  [57,]  47   5      1          3        1   130
#>  [58,]  46   3      1          0        1   119
#>  [59,]  49   4      0        180        0   117
#>  [60,]  47  10      0         10        1   140
#>  [61,]  48   2      1          0        0   184
#>  [62,]  50   4      1          1        0   125
#>  [63,]  46   9      1          9        1   122
#>  [64,]  50   7      0        180        1   110
#>  [65,]  49   2      0          2        0   105
#>  [66,]  49  15      1         11        1   160
#>  [67,]  47   2      0        180        0   150
#>  [68,]  49  23      0        179        1   112
#>  [69,]  46   6      1          0        1   156
#>  [70,]  52   2      0        180        1   170
#>  [71,]  50   4      0          4        1   100
#>  [72,]  51   3      1          2        0   113
#>  [73,]  50   1      1          0        0   150
#>  [74,]  50   9      0        180        0   130
#>  [75,]  49   7      1          4        1    90
#>  [76,]  47   8      0        180        0   160
#>  [77,]  47   6      0        180        1   162
#>  [78,]  51   8      0        180        1   140
#>  [79,]  52   2      0        180        0   155
#>  [80,]  46   3      0        180        1   120
#>  [81,]  48   7      1          0        1   110
#>  [82,]  48  17      1         10        0   111
#>  [83,]  47   2      1          1        0   110
#>  [84,]  52   4      1          4        0   152
#>  [85,]  49   9      1          3        0   102
#>  [86,]  53   5      0        180        1   140
#>  [87,]  54  17      1         12        1   102
#>  [88,]  53   5      0         77        0   159
#>  [89,]  53   7      1          0        0   199
#>  [90,]  50  10      1          6        0   122
#>  [91,]  50  14      1         13        0   170
#>  [92,]  48   3      1          2        0   150
#>  [93,]  51  25      1          1        0   202
#>  [94,]  52  14      1          7        1   200
#>  [95,]  48   6      0        180        0   160
#>  [96,]  53   4      1          0        1   156
#>  [97,]  51  13      0         99        1   160
#>  [98,]  49  16      0         16        0   125
#>  [99,]  55   3      1          1        0   150
#> [100,]  54  23      1         10        0   131
#> [101,]  55   6      1          2        1   114
#> [102,]  54   9      1          1        0   130
#> [103,]  52   4      0        180        1   180
#> [104,]  51  13      1         11        0   145
#> [105,]  54   4      1          0        1   121
#> [106,]  50   3      0        174        1   153
#> [107,]  55  28      1         13        1   160
#> [108,]  49   6      1          0        1   130
#> [109,]  50   7      1          1        0   156
#> [110,]  50   7      1          0        1   127
#> [111,]  56   4      1          1        1   130
#> [112,]  52   5      0        175        1   117
#> [113,]  55   1      0        180        0   127
#> [114,]  54   1      0        180        0   162
#> [115,]  54   7      1          0        1   100
#> [116,]  56   3      0        180        1   193
#> [117,]  56   2      0        180        0   132
#> [118,]  52   8      0        180        0   119
#> [119,]  53  18      1          9        1   150
#> [120,]  54   3      0        180        1   180
#> [121,]  55   6      0        180        0   170
#> [122,]  52  16      0         16        0   152
#> [123,]  53  10      1          9        0   172
#> [124,]  52  16      1         14        0   170
#> [125,]  53   4      0        180        1   150
#> [126,]  55   6      0        180        1   100
#> [127,]  55   6      1          5        1   138
#> [128,]  54  12      1          0        1   190
#> [129,]  54   3      0        180        0   128
#> [130,]  56   3      0          8        1   139
#> [131,]  55   1      0          2        0   130
#> [132,]  57   3      0          3        0   120
#> [133,]  57   5      1          3        1   138
#> [134,]  57   1      0          1        1   100
#> [135,]  55  11      1          7        0   104
#> [136,]  56  14      1         11        0   130
#> [137,]  53   3      1          0        1   200
#> [138,]  57  10      0        180        1   170
#> [139,]  58   8      0          8        1   130
#> [140,]  54   5      0        180        1   108
#> [141,]  55   3      1          1        1   156
#> [142,]  53  21      1         13        1   130
#> [143,]  57   4      0        180        1   119
#> [144,]  58   6      1          0        1    90
#> [145,]  54  17      1          8        1   227
#> [146,]  55   9      1          2        1   147
#> [147,]  54  23      1          8        0   120
#> [148,]  57   4      1          2        1   185
#> [149,]  53   4      0        147        1   145
#> [150,]  53   7      1          0        1   120
#> [151,]  57  11      1         10        1   129
#> [152,]  55   3      1          2        0   140
#> [153,]  55   5      0          5        1   131
#> [154,]  54   7      1          0        1   141
#> [155,]  59  15      1         10        0   140
#> [156,]  58   9      1          0        1   180
#> [157,]  58   1      1          1        1   200
#> [158,]  56   7      1          5        1   120
#> [159,]  55   2      0          2        0   106
#> [160,]  57   1      0        180        0   148
#> [161,]  60  11      1          9        0   106
#> [162,]  59   3      0        180        0   120
#> [163,]  57   2      0          2        1   120
#> [164,]  55   5      1          0        1   160
#> [165,]  57  10      1          9        0   103
#> [166,]  59   6      1          0        1   140
#> [167,]  59   5      0        180        1   155
#> [168,]  59   4      1          0        1   152
#> [169,]  58  26      1          0        1   189
#> [170,]  61   9      0          9        1   160
#> [171,]  58   4      1          3        0   120
#> [172,]  60   0      1          0        1    80
#> [173,]  59   2      1          1        0   140
#> [174,]  58   8      0        161        1   140
#> [175,]  58  14      1          6        0   190
#> [176,]  61   9      1          8        0   150
#> [177,]  61   3      1          2        1   102
#> [178,]  58   1      0          1        1   100
#> [179,]  61  20      1         13        0   130
#> [180,]  57  13      1         10        0   110
#> [181,]  58  10      0         10        1   150
#> [182,]  57  11      0        180        1   150
#> [183,]  61   3      0         17        0   143
#> [184,]  56  14      0         45        0   130
#> [185,]  57   3      1          2        0   120
#> [186,]  58  19      1         13        1   140
#> [187,]  56  18      1         11        1   165
#> [188,]  59   9      1          0        1    80
#> [189,]  58  11      0        172        1   135
#> [190,]  60  12      1          0        1   114
#> [191,]  61  13      1         12        1   130
#> [192,]  59  11      1          8        1   190
#> [193,]  59   5      1          2        0   182
#> [194,]  59  10      0        180        0   160
#> [195,]  61  13      0         13        0   210
#> [196,]  58   8      1          5        0   152
#> [197,]  62  10      1          0        1   153
#> [198,]  62   7      1          2        1   180
#> [199,]  61  18      0        170        0   140
#> [200,]  61  28      1          7        0   133
#> [201,]  57   7      0        169        0   180
#> [202,]  61   7      0          7        1   150
#> [203,]  60   7      0          7        0   147
#> [204,]  61   6      0          6        0   134
#> [205,]  59  13      1          2        0   198
#> [206,]  57  12      1          9        1   120
#> [207,]  62   4      1          0        0   160
#> [208,]  60  17      1          8        1   140
#> [209,]  58   3      1          0        1   146
#> [210,]  62   4      1          3        0   173
#> [211,]  58   2      0         30        0   202
#> [212,]  59   1      0        180        0   155
#> [213,]  59  16      1          9        1   133
#> [214,]  63   6      0         28        1   120
#> [215,]  57  18      1          9        1    93
#> [216,]  58   7      0        180        1   150
#> [217,]  62   3      0        180        1   105
#> [218,]  63  15      1         10        1   126
#> [219,]  61   9      1          9        1   150
#> [220,]  58   9      1          9        0   110
#> [221,]  62   7      0          7        0   150
#> [222,]  59   1      0         22        1   162
#> [223,]  59   4      0        180        0   196
#> [224,]  60   7      0          7        0   140
#> [225,]  59   5      1          1        0   148
#> [226,]  60   7      1          1        1    90
#> [227,]  65  13      0        180        1   100
#> [228,]  63   1      0          1        0   162
#> [229,]  62   6      0        180        0   170
#> [230,]  61  15      1         13        0   170
#> [231,]  62   6      0          6        0   120
#> [232,]  63  12      1         10        0   200
#> [233,]  59  10      0        180        1   130
#> [234,]  60   8      0         17        1   130
#> [235,]  61   6      1          1        1   117
#> [236,]  64  12      1         11        0   160
#> [237,]  66   1      1          0        1   120
#> [238,]  64   6      1          0        1   140
#> [239,]  65  36      1         11        0   140
#> [240,]  64  32      1          9        1   160
#> [241,]  65   8      1          0        0   168
#> [242,]  65  10      1          8        1   120
#> [243,]  64   0      0          0        1    90
#> [244,]  60   6      0        180        0   130
#> [245,]  64   9      0        180        0   150
#> [246,]  61   4      0        180        1   113
#> [247,]  65   3      0        180        1   190
#> [248,]  63  16      1          7        1   110
#> [249,]  66   6      1          1        1   130
#> [250,]  63  12      0         12        1   150
#> [251,]  62   3      1          1        1   199
#> [252,]  65   6      0          9        0   112
#> [253,]  63   5      1          4        0   170
#> [254,]  63   2      1          1        0   180
#> [255,]  62  13      1         11        0   180
#> [256,]  64   2      0          2        0   201
#> [257,]  66  18      1          5        0   142
#> [258,]  62   9      0        180        0   145
#> [259,]  61  15      1         10        0   130
#> [260,]  63   3      1          2        0   120
#> [261,]  63   2      1          0        0   140
#> [262,]  64  19      1          8        1   160
#> [263,]  65  15      1         11        1   160
#> [264,]  68   5      1          4        1   150
#> [265,]  64  13      1         12        1   150
#> [266,]  64   6      1          0        1   125
#> [267,]  66  13      1          0        0   118
#> [268,]  64  14      1         13        1   150
#> [269,]  65   3      0          3        0   105
#> [270,]  67   4      1          3        0   130
#> [271,]  66   3      1          0        1   135
#> [272,]  65   2      1          1        1   170
#> [273,]  64  10      1          9        1   110
#> [274,]  67   8      1          1        1   130
#> [275,]  63  10      0         16        1   160
#> [276,]  66  14      0        180        0   130
#> [277,]  64   1      0          1        1   120
#> [278,]  68  18      0        180        1   260
#> [279,]  63   1      1          0        1   155
#> [280,]  63  10      0         18        1   130
#> [281,]  67  11      0         11        0   150
#> [282,]  68  11      0        180        0   160
#> [283,]  68  14      0         79        0   172
#> [284,]  66  12      1         10        1   150
#> [285,]  65  15      1         12        1   150
#> [286,]  66  11      1          0        0   100
#> [287,]  65   4      1          2        1   145
#> [288,]  69  12      0         15        1   140
#> [289,]  63   2      0        180        0   150
#> [290,]  65  11      1          6        0   130
#> [291,]  69   6      0        180        1   100
#> [292,]  66   9      1          8        0   130
#> [293,]  68  14      1         13        1   140
#> [294,]  65   8      1          0        1    90
#> [295,]  66   3      0          3        1   138
#> [296,]  69   1      1          0        0   170
#> [297,]  67   1      0        180        1   160
#> [298,]  65   1      1          0        0   133
#> [299,]  67   7      1          4        1   130
#> [300,]  63   2      1          0        0    99
#> [301,]  67   2      0        180        0   184
#> [302,]  65   6      0          6        0    80
#> [303,]  65  10      1          1        1   148
#> [304,]  67  12      1         12        0   160
#> [305,]  69   6      0         99        1   140
#> [306,]  70  15      1         12        1   132
#> [307,]  64   4      0        180        1   140
#> [308,]  64   0      1          0        1   118
#> [309,]  66   7      1          5        1   131
#> [310,]  66   4      0        180        0   177
#> [311,]  68   4      1          0        1   160
#> [312,]  69  17      1         10        0   140
#> [313,]  69   8      0         93        0   140
#> [314,]  66   6      0        180        0   140
#> [315,]  68  18      1          0        1   160
#> [316,]  65   6      0        101        1   115
#> [317,]  68   4      0          4        1   190
#> [318,]  71   3      0          5        0   112
#> [319,]  70   7      1          0        1   190
#> [320,]  68   7      0        150        0   210
#> [321,]  71  20      1          0        1   160
#> [322,]  66   9      1          3        1   151
#> [323,]  70   4      1          0        1   180
#> [324,]  69   8      0        180        1   153
#> [325,]  70  14      0        171        0   166
#> [326,]  67  10      1          9        0   200
#> [327,]  67   6      1          4        0   130
#> [328,]  68  18      1         14        1   170
#> [329,]  69   0      0          0        1   148
#> [330,]  65   2      0        180        0   130
#> [331,]  68   7      1          0        1   150
#> [332,]  69   3      1          2        0   151
#> [333,]  67  14      1         13        0   130
#> [334,]  71   7      0          7        0   230
#> [335,]  69   5      0          5        1   142
#> [336,]  71   3      0        103        0   133
#> [337,]  69   3      0          3        1   130
#> [338,]  67   1      0         36        1   104
#> [339,]  67   5      0          5        0   130
#> [340,]  68   6      0        180        0   145
#> [341,]  69   6      1          4        1   174
#> [342,]  72   7      0          7        1   110
#> [343,]  69   8      1          7        1   108
#> [344,]  67   3      0        180        0   110
#> [345,]  66   2      1          1        0   123
#> [346,]  69  19      0        180        0   130
#> [347,]  68  18      0         18        1   100
#> [348,]  67  14      0        172        1   140
#> [349,]  69  11      1          0        1   120
#> [350,]  66   2      0        180        0   130
#> [351,]  67   7      1          4        0   122
#> [352,]  70   9      0        180        1   142
#> [353,]  68   3      0         19        0   135
#> [354,]  67  12      1          8        0   120
#> [355,]  69   1      0          1        1   110
#> [356,]  67   4      0         60        1   136
#> [357,]  69   5      0         76        0   120
#> [358,]  67   8      1          0        1   130
#> [359,]  72  13      1         11        1   195
#> [360,]  70   7      0          7        0   102
#> [361,]  71   6      0          9        0   120
#> [362,]  69  10      1          6        1   120
#> [363,]  70  11      0        180        1   210
#> [364,]  72  19      1          8        0   120
#> [365,]  72  12      1         10        0   170
#> [366,]  67   8      0        180        1   170
#> [367,]  67   9      0        180        0   158
#> [368,]  70   5      0        180        0   150
#> [369,]  67   4      1          1        0   134
#> [370,]  72   6      1          5        0   115
#> [371,]  71   1      0        173        1   188
#> [372,]  69   3      0        180        0   220
#> [373,]  72   5      0         28        0   120
#> [374,]  71   5      0        180        0   191
#> [375,]  73   6      0        180        1   117
#> [376,]  69   8      1          1        0   164
#> [377,]  68   7      0        180        1   130
#> [378,]  70   4      0        180        0   180
#> [379,]  69   1      1          0        0   155
#> [380,]  73   6      1          0        1   270
#> [381,]  72   8      1          1        1   150
#> [382,]  68  15      1         13        1   130
#> [383,]  70   3      0          3        1   159
#> [384,]  72   6      0        180        1   130
#> [385,]  69   2      1          0        1   110
#> [386,]  71   3      1          1        0   150
#> [387,]  71  15      1         11        0   165
#> [388,]  74  20      0         20        1   180
#> [389,]  71  20      1         10        0   140
#> [390,]  73   3      1          0        1   136
#> [391,]  71  17      1         11        0   160
#> [392,]  71   8      1          7        0   149
#> [393,]  71   3      1          2        1   190
#> [394,]  73  10      1          8        0   106
#> [395,]  70  26      1         11        1   120
#> [396,]  74   4      0          4        0   120
#> [397,]  70   3      0        180        1   154
#> [398,]  72  15      1          0        1   150
#> [399,]  71   7      1          2        0   143
#> [400,]  74   3      0          3        1   150
#> [401,]  73  17      1         11        0   140
#> [402,]  71  13      1          8        0   121
#> [403,]  69   2      1          1        1    80
#> [404,]  71  14      1         13        1   170
#> [405,]  74   7      1          0        1   117
#> [406,]  72  10      1          8        1   153
#> [407,]  72  15      1         13        0   156
#> [408,]  70   8      0          8        0   120
#> [409,]  71  10      1          9        1   120
#> [410,]  75   1      0          1        0   133
#> [411,]  75   2      1          1        0   145
#> [412,]  73  10      1          9        1   146
#> [413,]  72  10      1          9        1   160
#> [414,]  73  10      1         10        1   120
#> [415,]  74  15      1          9        1   179
#> [416,]  75   9      1          7        0   140
#> [417,]  75  13      1          1        1   130
#> [418,]  71   4      0          4        0   134
#> [419,]  72  15      1         12        1   120
#> [420,]  73  10      1          8        0   120
#> [421,]  72   1      1          1        0   168
#> [422,]  72   7      0         57        1   145
#> [423,]  73  10      0        180        0   162
#> [424,]  70   3      0          3        0   150
#> [425,]  73   5      1          3        1   112
#> [426,]  76  25      1         12        1   170
#> [427,]  73  12      1         12        1   140
#> [428,]  71   3      1          0        0   144
#> [429,]  73   5      0        180        0   126
#> [430,]  73   4      0        180        0   124
#> [431,]  76   3      1          0        1   120
#> [432,]  72   5      0        180        0   154
#> [433,]  72   3      0        180        0   160
#> [434,]  76   5      0          5        1   130
#> [435,]  77  11      0         11        1   150
#> [436,]  77   4      0          4        0   185
#> [437,]  72   7      1          2        0   142
#> [438,]  71  16      0        180        0   140
#> [439,]  73  10      1         10        0   124
#> [440,]  74   3      0          3        1   128
#> [441,]  76   1      0        180        0   114
#> [442,]  74   2      1          1        0   140
#> [443,]  76   8      1          0        1   141
#> [444,]  73   6      0          6        1   114
#> [445,]  74   2      0        180        0   190
#> [446,]  72   4      0         85        1   120
#> [447,]  72   4      1          3        0   160
#> [448,]  76  17      1          0        1   200
#> [449,]  76  13      1         10        0   110
#> [450,]  75   4      1          0        1   122
#> [451,]  75   7      0          7        0   190
#> [452,]  75   0      0          0        1   130
#> [453,]  73  13      1         11        0   195
#> [454,]  75  12      0         12        1   160
#> [455,]  76  13      1          8        1   148
#> [456,]  75   4      1          2        1   188
#> [457,]  76   4      0          4        1   155
#> [458,]  75   1      0          1        1   125
#> [459,]  73   1      0         52        1   105
#> [460,]  73   0      0        180        0   156
#> [461,]  72   5      0        180        0   120
#> [462,]  76  44      1         10        0   105
#> [463,]  74  10      1          0        1   135
#> [464,]  75   9      0        180        1   140
#> [465,]  73  33      1         12        1   175
#> [466,]  77  12      1          9        1   100
#> [467,]  73  10      1          9        0   146
#> [468,]  77   1      1          0        1    90
#> [469,]  78   5      1          0        1   170
#> [470,]  73   7      1          0        0   174
#> [471,]  74   6      0         79        1   140
#> [472,]  74   9      1          8        0   200
#> [473,]  75   6      0        180        0   150
#> [474,]  79  10      1          8        0   190
#> [475,]  77   3      0        180        0   110
#> [476,]  76  29      0         47        0    90
#> [477,]  73  11      1          2        1   110
#> [478,]  74   2      0        180        0   100
#> [479,]  78   7      0          7        1   133
#> [480,]  78  32      1          9        1   198
#> [481,]  79   6      0        180        0   170
#> [482,]  78  13      1          5        0   130
#> [483,]  78  15      0        180        1   270
#> [484,]  80   8      0          8        1   120
#> [485,]  74  10      1          8        0   135
#> [486,]  76   1      0          1        1    83
#> [487,]  79   4      0         80        0   145
#> [488,]  78  12      1          9        0   150
#> [489,]  75   4      1          0        0   212
#> [490,]  77   2      1          0        1   143
#> [491,]  78  10      0        180        1   130
#> [492,]  76  11      1          0        0   120
#> [493,]  75  11      1          4        0   162
#> [494,]  75   3      0          3        0     0
#> [495,]  77  24      0         24        1   160
#> [496,]  80   9      0         23        1   128
#> [497,]  80   6      0          6        1   150
#> [498,]  78   6      1          0        1   240
#> [499,]  76   3      1          0        1   140
#> [500,]  78  11      1          1        1   140
#> [501,]  79  11      0        180        0   160
#> [502,]  81   1      0          1        0   130
#> [503,]  78  11      1          8        1   118
#> [504,]  76   4      0          4        1   160
#> [505,]  79   4      0          4        1   125
#> [506,]  76  10      1          8        0   180
#> [507,]  77   6      0          6        1   107
#> [508,]  76   1      0          1        1   140
#> [509,]  77  31      1          3        1   161
#> [510,]  79   3      0          3        0   120
#> [511,]  77   7      0        180        1   170
#> [512,]  77   6      0          6        1   144
#> [513,]  79   4      1          0        1   120
#> [514,]  81   1      0        180        0   120
#> [515,]  77   9      1          4        0   141
#> [516,]  82   5      0          8        1   120
#> [517,]  78   4      0         59        1   112
#> [518,]  79  10      0         10        1   120
#> [519,]  80  15      1          0        1    90
#> [520,]  81   4      1          2        1   126
#> [521,]  79  28      0        164        0   100
#> [522,]  80   9      0        118        1   186
#> [523,]  80   6      0        173        1   160
#> [524,]  78  32      0        180        1   130
#> [525,]  79   1      0         37        1   140
#> [526,]  81   3      0        180        0   184
#> [527,]  81   2      0        175        0   172
#> [528,]  80   5      1          1        1   108
#> [529,]  78   4      0        180        0   175
#> [530,]  79   3      0          3        1   101
#> [531,]  78  26      1          5        0   194
#> [532,]  76   1      0        166        0   131
#> [533,]  81   4      1          1        1   104
#> [534,]  78  20      1          0        1   109
#> [535,]  80   1      0          1        0   100
#> [536,]  78   3      1          1        1   152
#> [537,]  77  10      1          8        1   130
#> [538,]  82   3      1          1        1   144
#> [539,]  77   5      0         85        0   188
#> [540,]  79   6      0          6        0   152
#> [541,]  80   6      1          0        1   119
#> [542,]  78   2      0        180        0   148
#> [543,]  80   5      0          5        1   130
#> [544,]  82  21      1          2        0   155
#> [545,]  84  22      1         10        0   180
#> [546,]  79   4      0          4        1   121
#> [547,]  80   6      0          6        1   110
#> [548,]  83   9      1          5        1   170
#> [549,]  82   5      0        180        0   110
#> [550,]  83   5      0        180        0   148
#> [551,]  79   7      1          6        0   130
#> [552,]  83   4      0        103        0    97
#> [553,]  81  11      1          8        0   160
#> [554,]  81   5      0        177        0    41
#> [555,]  80  11      1          8        0   170
#> [556,]  78  23      1         10        1   145
#> [557,]  79   4      0          4        1   183
#> [558,]  78   9      1          4        1   120
#> [559,]  79   1      0        180        1   170
#> [560,]  81  15      0        180        1   140
#> [561,]  80   7      1          0        1   146
#> [562,]  84   5      1          1        1    85
#> [563,]  81  20      1          9        0   170
#> [564,]  83   8      0          8        0   115
#> [565,]  81  16      0         16        1   110
#> [566,]  80   6      1          0        1   150
#> [567,]  81   8      0        180        0   146
#> [568,]  80   8      1          7        0   160
#> [569,]  83   2      0          2        1   155
#> [570,]  84   5      0        180        1   203
#> [571,]  84   4      0          4        1    85
#> [572,]  81   1      0          1        1   150
#> [573,]  84   1      0         38        1   205
#> [574,]  83   3      0        180        0   174
#> [575,]  81   4      0         90        1   138
#> [576,]  79   9      1          8        0   150
#> [577,]  79   4      0          4        1    60
#> [578,]  80   6      0         71        1   189
#> [579,]  83   1      0          1        1   100
#> [580,]  80  30      1         13        0   220
#> [581,]  81  14      1         12        1   128
#> [582,]  85   9      1          6        1   160
#> [583,]  84  15      1         13        1   110
#> [584,]  81   1      0          1        1   145
#> [585,]  81  12      0         12        1   163
#> [586,]  82  16      1          8        0   103
#> [587,]  82   5      1          0        1   146
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  82   3      1          2        0   130
#> [592,]  80   2      0         88        0   135
#> [593,]  86   8      0          8        1   132
#> [594,]  81  16      1          9        0   180
#> [595,]  81  13      0        180        0   152
#> [596,]  85   3      0          3        1   118
#> [597,]  81   4      0        180        0   160
#> [598,]  83   9      0        180        1   149
#> [599,]  82   1      0        180        1   193
#> [600,]  83   4      0          4        0   130
#> [601,]  87   2      0          5        1   137
#> [602,]  86  12      1          0        1   132
#> [603,]  84  16      0         70        1   150
#> [604,]  83  19      0         43        0   150
#> [605,]  84   3      1          2        0   125
#> [606,]  83  10      1          0        1   190
#> [607,]  86   2      0        180        1   169
#> [608,]  83  13      1         12        0   170
#> [609,]  87   2      0        180        0   113
#> [610,]  84   9      0         92        1   110
#> [611,]  86   4      0         38        1   122
#> [612,]  86  13      0        177        0   163
#> [613,]  86   6      0          6        1   117
#> [614,]  84  13      0         62        1   100
#> [615,]  88   4      0          4        0   100
#> [616,]  83  20      1          3        1   150
#> [617,]  85  22      0         22        1   184
#> [618,]  83   9      0         65        1   150
#> [619,]  86   9      1          7        1   142
#> [620,]  87   2      0        180        1   130
#> [621,]  86   6      0         46        0   173
#> [622,]  88   3      0        115        0   110
#> [623,]  88   2      0        180        1    68
#> [624,]  83   3      0          3        1   130
#> [625,]  87   8      0          8        1   157
#> [626,]  86  15      1          8        1   109
#> [627,]  88   4      0          4        0    86
#> [628,]  89   5      0        119        1   140
#> [629,]  87   6      0        180        1   110
#> [630,]  84   8      0        180        1   119
#> [631,]  85   8      0          8        1   136
#> [632,]  84   2      0        110        1   174
#> [633,]  87  29      0         29        1    97
#> [634,]  89  10      0         46        1   170
#> [635,]  90  14      0         14        1   100
#> [636,]  88   1      0          1        0   135
#> [637,]  86   4      0        180        1   145
#> [638,]  88   7      0         24        0   119
#> [639,]  88   8      0         50        1   154
#> [640,]  90  11      1         10        1   186
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  87  43      0        178        1   130
#> [645,]  87   5      0         36        1   150
#> [646,]  90   5      1          0        1   125
#> [647,]  88   3      1          2        0   159
#> [648,]  91   3      0         33        1   137
#> [649,]  88   5      0        158        0   100
#> [650,]  87   7      0         74        1   105
#> [651,]  89   2      0        168        0   118
#> [652,]  89  52      0         52        1   130
#> [653,]  89  14      0        180        1    84
#> [654,]  90  18      0        180        0   188
#> [655,]  91   4      1          0        1   120
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  93   8      0        179        1   110
#> [660,]  94   8      0          8        1   142
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+
#>  [11] 115.0  180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+
#>  [21]   2.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+   2.0  179.0+ 180.0+ 179.0+ 180.0+ 180.0+
#>  [71]   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81]   7.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+  85.0  180.0+ 166.0+  99.0   16.0+ 180.0+ 152.0+
#> [101]   6.0+ 180.0+ 180.0+  13.0+ 180.0+ 174.0+  28.0    6.0+ 180.0+ 180.0+
#> [111] 180.0+ 175.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [121] 180.0+  16.0+ 180.0+  16.0  180.0+ 180.0+ 180.0+  12.0+ 180.0+   8.0 
#> [131]   2.0    3.0+ 140.0    1.0  180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+
#> [141] 180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0  180.0+   4.0+ 147.0+ 180.0+
#> [151] 180.0+ 180.0+   5.0+ 180.0+ 180.0+   9.0+   1.0  180.0+   2.0+ 180.0+
#> [161] 180.0+ 180.0+   2.0  180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+   9.0+
#> [171] 180.0+   0.5  180.0+ 161.0+ 171.0+ 180.0+   3.0    1.0  180.0+ 180.0+
#> [181]  10.0+ 180.0+  17.0   45.0    3.0+  19.0  180.0+   9.0+ 172.0+ 172.0+
#> [191] 180.0+ 180.0+ 180.0+ 180.0+  13.0+   8.0+ 180.0+ 180.0+ 170.0   94.0 
#> [201] 169.0    7.0    7.0+   6.0  180.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+
#> [211]  30.0  180.0+ 180.0+  28.0   18.0  180.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [221]   7.0+  22.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [231]   6.0+ 180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+  36.0  180.0+
#> [241] 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0 
#> [251] 180.0+   9.0  180.0+ 180.0+ 180.0+   2.0+  18.0+ 180.0+ 180.0+   3.0+
#> [261]   2.0+ 103.0  180.0+   5.0+  13.0  180.0+ 166.0+  14.0+   3.0  180.0+
#> [271]   3.0+ 175.0+ 180.0+   8.0   16.0  180.0+   1.0  180.0+   1.0+  18.0 
#> [281]  11.0+ 180.0+  79.0   80.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+
#> [291] 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0  180.0+ 180.0+ 180.0+ 180.0+
#> [301] 180.0+   6.0  180.0+  12.0   99.0  180.0+ 180.0+   0.5    7.0+ 180.0+
#> [311] 180.0+ 180.0+  93.0  180.0+  18.0+ 101.0    4.0    5.0    7.0+ 150.0 
#> [321] 180.0+ 180.0+ 180.0+ 180.0+ 171.0  174.0+   6.0  180.0+   0.5  180.0+
#> [331] 180.0+ 180.0+ 180.0+   7.0+   5.0+ 103.0    3.0+  36.0    5.0+ 180.0+
#> [341]  97.0    7.0    8.0+ 180.0+   2.0+ 180.0+  18.0  172.0+ 180.0+ 180.0+
#> [351]   7.0  180.0+  19.0  180.0+   1.0   60.0   76.0  180.0+ 132.0    7.0+
#> [361]   9.0  180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+  76.0  180.0+
#> [371] 173.0+ 180.0+  28.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0 
#> [381] 180.0+  15.0    3.0+ 180.0+   2.0    3.0+ 180.0+  20.0   20.0  180.0+
#> [391] 180.0+   8.0    3.0   87.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+   3.0 
#> [401] 180.0+ 175.0    2.0   14.0+ 180.0+  10.0+ 180.0+   8.0+ 179.0+   1.0 
#> [411] 180.0+ 180.0+ 159.0   15.0  180.0+ 180.0+  13.0    4.0+ 180.0+  10.0 
#> [421]   1.0   57.0  180.0+   3.0+   5.0  180.0+  12.0  180.0+ 180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+   5.0   11.0+   4.0+   7.0  180.0+  10.0    3.0 
#> [441] 180.0+ 180.0+ 180.0+   6.0  180.0+  85.0  180.0+  17.0+ 174.0+   4.0 
#> [451]   7.0    0.5  180.0+  12.0  180.0+  46.0    4.0    1.0   52.0  180.0+
#> [461] 180.0+ 180.0+ 180.0+ 180.0+  33.0  180.0+ 180.0+   1.0  180.0+   7.0+
#> [471]  79.0  168.0+ 180.0+ 180.0+ 180.0+  47.0   11.0  180.0+   7.0   32.0 
#> [481] 180.0+ 172.0  180.0+   8.0  180.0+   1.0   80.0  180.0+   4.0+   2.0 
#> [491] 180.0+  11.0  152.0+   3.0   24.0   23.0    6.0  180.0+   3.0+ 180.0+
#> [501] 180.0+   1.0   11.0    4.0    4.0   10.0+   6.0    1.0  171.0    3.0 
#> [511] 180.0+   6.0  138.0  180.0+  71.0    8.0   59.0   10.0+ 180.0+  93.0 
#> [521] 164.0  118.0  173.0  180.0+  37.0  180.0+ 175.0+   5.0+ 180.0+   3.0 
#> [531] 171.0+ 166.0+  71.0   20.0+   1.0    3.0+  10.0  180.0+  85.0    6.0+
#> [541]   6.0  180.0+   5.0  180.0+ 180.0+   4.0    6.0    9.0+ 180.0+ 180.0+
#> [551] 180.0+ 103.0  180.0+ 177.0+ 169.0   70.0    4.0  180.0+ 180.0+ 180.0+
#> [561]   7.0+ 180.0+  20.0    8.0+  16.0  180.0+ 180.0+ 180.0+   2.0  180.0+
#> [571]   4.0    1.0   38.0  180.0+  90.0  180.0+   4.0   71.0    1.0   30.0 
#> [581] 180.0+ 180.0+ 180.0+   1.0   12.0   16.0+   5.0+   4.0+ 180.0+  77.0 
#> [591]   3.0   88.0    8.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   4.0+
#> [601]   5.0  180.0+  70.0   43.0  180.0+ 180.0+ 180.0+  13.0  180.0+  92.0 
#> [611]  38.0  177.0    6.0+  62.0    4.0+  20.0   22.0   65.0   11.0  180.0+
#> [621]  46.0  115.0  180.0+   3.0+   8.0+ 180.0+   4.0  119.0  180.0+ 180.0+
#> [631]   8.0  110.0   29.0   46.0   14.0    1.0+ 180.0+  24.0   50.0   11.0 
#> [641] 180.0+ 180.0+   4.0  178.0+  36.0   89.0   75.0   33.0  158.0   74.0 
#> [651] 168.0   52.0  180.0+ 180.0+   4.0  180.0+  50.0    1.0+ 179.0+   8.0+
#> [661]  12.0    8.0   26.0   53.0    0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8230551