Survival Logistic-Hazard Learner

Survival logistic hazard learner. Calls survivalmodels::loghaz() from package 'survivalmodels'.

Details

Custom nets can be used in this learner either using the survivalmodels::build_pytorch_net utility function or using torch via reticulate. The number of output channels depends on the number of discretised time-points, i.e. the parameters cuts or cutpoints.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.loghaz")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, survivalmodels, distr6, reticulate

Parameters

Id	Type	Default	Levels	Range

frac	numeric	0		\([0, 1]\)

cuts	integer	10		\([1, \infty)\)

cutpoints	untyped	-		-

scheme	character	equidistant	equidistant, quantiles	-

cut_min	numeric	0		\([0, \infty)\)

num_nodes	untyped	c(32L, 32L)		-

batch_norm	logical	TRUE	TRUE, FALSE	-

dropout	numeric	-		\([0, 1]\)

activation	character	relu	celu, elu, gelu, glu, hardshrink, hardsigmoid, hardswish, hardtanh, relu6, leakyrelu, ...	-

custom_net	untyped	-		-

device	untyped	-		-

optimizer	character	adam	adadelta, adagrad, adam, adamax, adamw, asgd, rmsprop, rprop, sgd, sparse_adam	-

rho	numeric	0.9		\((-\infty, \infty)\)

eps	numeric	1e-08		\((-\infty, \infty)\)

lr	numeric	1		\((-\infty, \infty)\)

weight_decay	numeric	0		\((-\infty, \infty)\)

learning_rate	numeric	0.01		\((-\infty, \infty)\)

lr_decay	numeric	0		\((-\infty, \infty)\)

betas	untyped	c(0.9, 0.999)		-

amsgrad	logical	FALSE	TRUE, FALSE	-

lambd	numeric	1e-04		\([0, \infty)\)

alpha	numeric	0.75		\([0, \infty)\)

t0	numeric	1e+06		\((-\infty, \infty)\)

momentum	numeric	0		\((-\infty, \infty)\)

centered	logical	TRUE	TRUE, FALSE	-

etas	untyped	c(0.5, 1.2)		-

step_sizes	untyped	c(1e-06, 50)		-

dampening	numeric	0		\((-\infty, \infty)\)

nesterov	logical	FALSE	TRUE, FALSE	-

batch_size	integer	256		\((-\infty, \infty)\)

epochs	integer	1		\([1, \infty)\)

verbose	logical	TRUE	TRUE, FALSE	-

num_workers	integer	0		\((-\infty, \infty)\)

shuffle	logical	TRUE	TRUE, FALSE	-

best_weights	logical	FALSE	TRUE, FALSE	-

early_stopping	logical	FALSE	TRUE, FALSE	-

min_delta	numeric	0		\((-\infty, \infty)\)

patience	integer	10		\((-\infty, \infty)\)

interpolate	logical	FALSE	TRUE, FALSE	-

inter_scheme	character	const_hazard	const_hazard, const_pdf	-

sub	integer	10		\([1, \infty)\)

Installation

Package 'survivalmodels' is not on CRAN and has to be install from GitHub via remotes::install_github("RaphaelS1/survivalmodels").

References

Gensheimer, F M, Narasimhan, BA (2018). “Simple discrete-time survival model for neural networks.” arXiv.

Kvamme, Håvard, Borgan Ø, Scheel I (2019). “Time-to-event prediction with neural networks and Cox regression.” arXiv preprint arXiv:1907.00825.

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvLogisticHazard

Methods

Public methods

LearnerSurvLogisticHazard$new()

LearnerSurvLogisticHazard$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvLogisticHazard$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvLogisticHazard$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

lrn("surv.loghaz")
#> <LearnerSurvLogisticHazard:surv.loghaz>: Logistic-Hazard Learner
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, survivalmodels, distr6,
#>   reticulate
#> * Predict Types:  [crank], distr
#> * Feature Types: integer, numeric
#> * Properties: -