Classification Logistic Model Trees Learner

Calls RWeka::LMT from package RWeka.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.LMT")
lrn("classif.LMT")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “numeric”, “factor”, “ordered”

Required Packages: mlr3extralearners, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	list	-		\((-\infty, \infty)\)

na.action	list	-		\((-\infty, \infty)\)

B	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

R	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

C	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

P	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

I	integer	-		\([1, \infty)\)

M	integer	15		\([1, \infty)\)

W	numeric	0		\([0, 1]\)

A	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

doNotMakeSplitPointActualValue	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	list	NULL		\((-\infty, \infty)\)

Custom mlr3 defaults

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

References

Landwehr N, Hall M, Frank E (2005). Logistic Model Trees https://link.springer.com/content/pdf/10.1007/s10994-005-0466-3.pdf

Author

henrifnk

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLMT

Methods

Public methods

LearnerClassifLMT$new()

LearnerClassifLMT$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLMT$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLMT$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.