Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 M (0.5467626 0.4532374)  
#>   2) V11>=0.1791 83 15 M (0.8192771 0.1807229)  
#>     4) V9>=0.09935 76  9 M (0.8815789 0.1184211) *
#>     5) V9< 0.09935 7  1 R (0.1428571 0.8571429) *
#>   3) V11< 0.1791 56  8 R (0.1428571 0.8571429)  
#>     6) V20>=0.8321 14  6 M (0.5714286 0.4285714) *
#>     7) V20< 0.8321 42  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V12>=0.2382 75 18 M (0.76000000 0.24000000)  
#>      4) V51>=0.01125 55  6 M (0.89090909 0.10909091) *
#>      5) V51< 0.01125 20  8 R (0.40000000 0.60000000)  
#>       10) V51< 0.00445 8  0 M (1.00000000 0.00000000) *
#>       11) V51>=0.00445 12  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.2382 64 10 R (0.15625000 0.84375000)  
#>      6) V21>=0.8088 15  7 M (0.53333333 0.46666667) *
#>      7) V21< 0.8088 49  2 R (0.04081633 0.95918367) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 R (0.41007194 0.58992806)  
#>    2) V45>=0.19815 45 10 M (0.77777778 0.22222222)  
#>      4) V37< 0.44565 26  0 M (1.00000000 0.00000000) *
#>      5) V37>=0.44565 19  9 R (0.47368421 0.52631579) *
#>    3) V45< 0.19815 94 22 R (0.23404255 0.76595745)  
#>      6) V11>=0.24655 40 18 R (0.45000000 0.55000000)  
#>       12) V8< 0.17125 22  5 M (0.77272727 0.22727273) *
#>       13) V8>=0.17125 18  1 R (0.05555556 0.94444444) *
#>      7) V11< 0.24655 54  4 R (0.07407407 0.92592593) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V11>=0.19685 91 34 M (0.62637363 0.37362637)  
#>      4) V17< 0.513 48  6 M (0.87500000 0.12500000)  
#>        8) V10>=0.14705 40  0 M (1.00000000 0.00000000) *
#>        9) V10< 0.14705 8  2 R (0.25000000 0.75000000) *
#>      5) V17>=0.513 43 15 R (0.34883721 0.65116279)  
#>       10) V31< 0.39185 20  5 M (0.75000000 0.25000000)  
#>         20) V23>=0.78455 12  0 M (1.00000000 0.00000000) *
#>         21) V23< 0.78455 8  3 R (0.37500000 0.62500000) *
#>       11) V31>=0.39185 23  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.19685 48  6 R (0.12500000 0.87500000)  
#>      6) V2>=0.04655 7  3 M (0.57142857 0.42857143) *
#>      7) V2< 0.04655 41  2 R (0.04878049 0.95121951) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V47>=0.0633 106 39 M (0.63207547 0.36792453)  
#>      4) V36< 0.50435 71 13 M (0.81690141 0.18309859)  
#>        8) V51>=0.0131 49  0 M (1.00000000 0.00000000) *
#>        9) V51< 0.0131 22  9 R (0.40909091 0.59090909)  
#>         18) V23>=0.7934 10  1 M (0.90000000 0.10000000) *
#>         19) V23< 0.7934 12  0 R (0.00000000 1.00000000) *
#>      5) V36>=0.50435 35  9 R (0.25714286 0.74285714)  
#>       10) V45>=0.25765 13  4 M (0.69230769 0.30769231) *
#>       11) V45< 0.25765 22  0 R (0.00000000 1.00000000) *
#>    3) V47< 0.0633 33  3 R (0.09090909 0.90909091) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V10>=0.1909 68 13 M (0.80882353 0.19117647)  
#>      4) V51>=0.00975 55  5 M (0.90909091 0.09090909) *
#>      5) V51< 0.00975 13  5 R (0.38461538 0.61538462) *
#>    3) V10< 0.1909 71 20 R (0.28169014 0.71830986)  
#>      6) V5>=0.13985 8  0 M (1.00000000 0.00000000) *
#>      7) V5< 0.13985 63 12 R (0.19047619 0.80952381)  
#>       14) V54>=0.0151 14  6 M (0.57142857 0.42857143) *
#>       15) V54< 0.0151 49  4 R (0.08163265 0.91836735) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V11>=0.19425 88 30 M (0.65909091 0.34090909)  
#>      4) V31< 0.4599 54  9 M (0.83333333 0.16666667)  
#>        8) V47>=0.02235 47  2 M (0.95744681 0.04255319) *
#>        9) V47< 0.02235 7  0 R (0.00000000 1.00000000) *
#>      5) V31>=0.4599 34 13 R (0.38235294 0.61764706)  
#>       10) V28>=0.6853 15  3 M (0.80000000 0.20000000) *
#>       11) V28< 0.6853 19  1 R (0.05263158 0.94736842) *
#>    3) V11< 0.19425 51 10 R (0.19607843 0.80392157)  
#>      6) V54< 0.00205 7  0 M (1.00000000 0.00000000) *
#>      7) V54>=0.00205 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V45>=0.2726 27  1 M (0.96296296 0.03703704) *
#>    3) V45< 0.2726 112 39 R (0.34821429 0.65178571)  
#>      6) V19>=0.7717 38 10 M (0.73684211 0.26315789)  
#>       12) V35>=0.25335 25  0 M (1.00000000 0.00000000) *
#>       13) V35< 0.25335 13  3 R (0.23076923 0.76923077) *
#>      7) V19< 0.7717 74 11 R (0.14864865 0.85135135)  
#>       14) V23>=0.88185 9  2 M (0.77777778 0.22222222) *
#>       15) V23< 0.88185 65  4 R (0.06153846 0.93846154) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V32< 0.39855 74 13 M (0.82432432 0.17567568)  
#>      4) V20>=0.49015 63  3 M (0.95238095 0.04761905) *
#>      5) V20< 0.49015 11  1 R (0.09090909 0.90909091) *
#>    3) V32>=0.39855 65 21 R (0.32307692 0.67692308)  
#>      6) V48>=0.08065 31 12 M (0.61290323 0.38709677)  
#>       12) V24>=0.611 16  0 M (1.00000000 0.00000000) *
#>       13) V24< 0.611 15  3 R (0.20000000 0.80000000) *
#>      7) V48< 0.08065 34  2 R (0.05882353 0.94117647) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V12>=0.21215 87 19 M (0.78160920 0.21839080)  
#>     4) V5>=0.01235 79 11 M (0.86075949 0.13924051)  
#>       8) V22< 0.86695 68  3 M (0.95588235 0.04411765) *
#>       9) V22>=0.86695 11  3 R (0.27272727 0.72727273) *
#>     5) V5< 0.01235 8  0 R (0.00000000 1.00000000) *
#>   3) V12< 0.21215 52 13 R (0.25000000 0.75000000)  
#>     6) V4>=0.0522 19  6 M (0.68421053 0.31578947) *
#>     7) V4< 0.0522 33  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.7133763 0.6885904 0.7864435 0.8335321 0.5986586 0.9781179
#>  [8] 0.6135867 0.8018062 0.6540283
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.4321907 5.9472803
#>   [2,] 1.4471188 5.9323522
#>   [3,] 1.9640180 5.4154531
#>   [4,] 2.6527171 4.7267539
#>   [5,] 1.4247073 5.9547637
#>   [6,] 2.0457774 5.3336936
#>   [7,] 2.2339969 5.1454741
#>   [8,] 1.9660633 5.4134078
#>   [9,] 0.8018062 6.5776649
#>  [10,] 2.2422780 5.1371930
#>  [11,] 0.9781179 6.4013531
#>  [12,] 2.8251721 4.5542989
#>  [13,] 2.7504615 4.6290096
#>  [14,] 2.7673260 4.6121450
#>  [15,] 2.2335623 5.1459087
#>  [16,] 0.0000000 7.3794710
#>  [17,] 0.7864435 6.5930275
#>  [18,] 2.0694212 5.3100498
#>  [19,] 2.2009368 5.1785343
#>  [20,] 1.3999214 5.9795496
#>  [21,] 0.9781179 6.4013531
#>  [22,] 1.4000302 5.9794408
#>  [23,] 0.6135867 6.7658844
#>  [24,] 0.0000000 7.3794710
#>  [25,] 0.0000000 7.3794710
#>  [26,] 0.5986586 6.7808124
#>  [27,] 0.5986586 6.7808124
#>  [28,] 2.3455225 5.0339485
#>  [29,] 1.4471188 5.9323522
#>  [30,] 2.4185897 4.9608813
#>  [31,] 2.9916260 4.3878451
#>  [32,] 2.9029533 4.4765177
#>  [33,] 0.6540283 6.7254427
#>  [34,] 0.8335321 6.5459389
#>  [35,] 0.0000000 7.3794710
#>  [36,] 0.8018062 6.5776649
#>  [37,] 0.0000000 7.3794710
#>  [38,] 0.6135867 6.7658844
#>  [39,] 0.6135867 6.7658844
#>  [40,] 0.0000000 7.3794710
#>  [41,] 0.5986586 6.7808124
#>  [42,] 1.2526870 6.1267841
#>  [43,] 0.6135867 6.7658844
#>  [44,] 0.0000000 7.3794710
#>  [45,] 0.0000000 7.3794710
#>  [46,] 0.0000000 7.3794710
#>  [47,] 0.0000000 7.3794710
#>  [48,] 1.4247073 5.9547637
#>  [49,] 1.4247073 5.9547637
#>  [50,] 3.0027429 4.3767281
#>  [51,] 3.0600456 4.3194254
#>  [52,] 2.0233659 5.3561051
#>  [53,] 2.0539497 5.3255213
#>  [54,] 2.2017276 5.1777435
#>  [55,] 2.9421359 4.4373352
#>  [56,] 2.3434772 5.0359938
#>  [57,] 2.1435217 5.2359493
#>  [58,] 2.3434772 5.0359938
#>  [59,] 1.3099896 6.0694814
#>  [60,] 0.7113310 6.6681400
#>  [61,] 3.1299207 4.2495503
#>  [62,] 1.4875605 5.8919106
#>  [63,] 2.4028252 4.9766458
#>  [64,] 0.6540283 6.7254427
#>  [65,] 1.2526870 6.1267841
#>  [66,] 5.8642886 1.5151825
#>  [67,] 6.5930275 0.7864435
#>  [68,] 5.7594954 1.6199756
#>  [69,] 6.5930275 0.7864435
#>  [70,] 4.2463582 3.1331128
#>  [71,] 4.4474605 2.9320106
#>  [72,] 5.2210212 2.1584498
#>  [73,] 4.3873017 2.9921694
#>  [74,] 4.9263192 2.4531518
#>  [75,] 5.6149096 1.7645614
#>  [76,] 5.8919106 1.4875605
#>  [77,] 6.6681400 0.7113310
#>  [78,] 6.7254427 0.6540283
#>  [79,] 5.9044371 1.4750339
#>  [80,] 5.1459087 2.2335623
#>  [81,] 5.4381937 1.9412773
#>  [82,] 5.3254125 2.0540585
#>  [83,] 6.6908807 0.6885904
#>  [84,] 6.0772940 1.3021771
#>  [85,] 6.0772940 1.3021771
#>  [86,] 5.4232656 1.9562054
#>  [87,] 5.9044371 1.4750339
#>  [88,] 5.1459087 2.2335623
#>  [89,] 6.0368523 1.3426187
#>  [90,] 6.6908807 0.6885904
#>  [91,] 5.1756982 2.2037729
#>  [92,] 5.0014318 2.3780393
#>  [93,] 5.6900221 1.6894489
#>  [94,] 5.8796512 1.4998198
#>  [95,] 7.3794710 0.0000000
#>  [96,] 6.5930275 0.7864435
#>  [97,] 5.7912214 1.5882497
#>  [98,] 5.8796512 1.4998198
#>  [99,] 5.7877665 1.5917046
#> [100,] 5.4232656 1.9562054
#> [101,] 5.3807822 1.9986888
#> [102,] 4.3622227 3.0172484
#> [103,] 6.6908807 0.6885904
#> [104,] 5.8890745 1.4903966
#> [105,] 5.8890745 1.4903966
#> [106,] 6.6908807 0.6885904
#> [107,] 5.8890745 1.4903966
#> [108,] 6.0674361 1.3120349
#> [109,] 5.8026945 1.5767765
#> [110,] 5.1141041 2.2653669
#> [111,] 4.3665141 3.0129570
#> [112,] 5.3788457 2.0006253
#> [113,] 5.3057785 2.0736925
#> [114,] 5.1305461 2.2489250
#> [115,] 5.2189759 2.1604951
#> [116,] 6.7254427 0.6540283
#> [117,] 5.9790063 1.4004648
#> [118,] 5.8573485 1.5221225
#> [119,] 6.5776649 0.8018062
#> [120,] 5.9323522 1.4471188
#> [121,] 6.7254427 0.6540283
#> [122,] 6.5776649 0.8018062
#> [123,] 7.3794710 0.0000000
#> [124,] 5.9640782 1.4153929
#> [125,] 5.9640782 1.4153929
#> [126,] 5.2754878 2.1039833
#> [127,] 6.1672257 1.2122453
#> [128,] 5.4786353 1.9008357
#> [129,] 6.7658844 0.6135867
#> [130,] 6.7658844 0.6135867
#> [131,] 5.4786353 1.9008357
#> [132,] 6.7658844 0.6135867
#> [133,] 6.7658844 0.6135867
#> [134,] 6.7658844 0.6135867
#> [135,] 6.7658844 0.6135867
#> [136,] 5.4538494 1.9256216
#> [137,] 6.0525080 1.3269630
#> [138,] 5.3336936 2.0457774
#> [139,] 5.9640782 1.4153929
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.19407770 0.80592230
#>   [2,] 0.19610062 0.80389938
#>   [3,] 0.26614617 0.73385383
#>   [4,] 0.35947253 0.64052747
#>   [5,] 0.19306361 0.80693639
#>   [6,] 0.27722548 0.72277452
#>   [7,] 0.30273131 0.69726869
#>   [8,] 0.26642333 0.73357667
#>   [9,] 0.10865361 0.89134639
#>  [10,] 0.30385349 0.69614651
#>  [11,] 0.13254580 0.86745420
#>  [12,] 0.38284209 0.61715791
#>  [13,] 0.37271797 0.62728203
#>  [14,] 0.37500331 0.62499669
#>  [15,] 0.30267241 0.69732759
#>  [16,] 0.00000000 1.00000000
#>  [17,] 0.10657180 0.89342820
#>  [18,] 0.28042948 0.71957052
#>  [19,] 0.29825129 0.70174871
#>  [20,] 0.18970484 0.81029516
#>  [21,] 0.13254580 0.86745420
#>  [22,] 0.18971959 0.81028041
#>  [23,] 0.08314779 0.91685221
#>  [24,] 0.00000000 1.00000000
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.08112487 0.91887513
#>  [27,] 0.08112487 0.91887513
#>  [28,] 0.31784426 0.68215574
#>  [29,] 0.19610062 0.80389938
#>  [30,] 0.32774568 0.67225432
#>  [31,] 0.40539843 0.59460157
#>  [32,] 0.39338230 0.60661770
#>  [33,] 0.08862808 0.91137192
#>  [34,] 0.11295283 0.88704717
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.10865361 0.89134639
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.08314779 0.91685221
#>  [39,] 0.08314779 0.91685221
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.08112487 0.91887513
#>  [42,] 0.16975295 0.83024705
#>  [43,] 0.08314779 0.91685221
#>  [44,] 0.00000000 1.00000000
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.19306361 0.80693639
#>  [49,] 0.19306361 0.80693639
#>  [50,] 0.40690490 0.59309510
#>  [51,] 0.41467005 0.58532995
#>  [52,] 0.27418848 0.72581152
#>  [53,] 0.27833292 0.72166708
#>  [54,] 0.29835845 0.70164155
#>  [55,] 0.39869197 0.60130803
#>  [56,] 0.31756710 0.68243290
#>  [57,] 0.29047092 0.70952908
#>  [58,] 0.31756710 0.68243290
#>  [59,] 0.17751809 0.82248191
#>  [60,] 0.09639322 0.90360678
#>  [61,] 0.42413890 0.57586110
#>  [62,] 0.20158090 0.79841910
#>  [63,] 0.32560941 0.67439059
#>  [64,] 0.08862808 0.91137192
#>  [65,] 0.16975295 0.83024705
#>  [66,] 0.79467600 0.20532400
#>  [67,] 0.89342820 0.10657180
#>  [68,] 0.78047537 0.21952463
#>  [69,] 0.89342820 0.10657180
#>  [70,] 0.57542854 0.42457146
#>  [71,] 0.60268012 0.39731988
#>  [72,] 0.70750616 0.29249384
#>  [73,] 0.59452793 0.40547207
#>  [74,] 0.66757078 0.33242922
#>  [75,] 0.76088240 0.23911760
#>  [76,] 0.79841910 0.20158090
#>  [77,] 0.90360678 0.09639322
#>  [78,] 0.91137192 0.08862808
#>  [79,] 0.80011658 0.19988342
#>  [80,] 0.69732759 0.30267241
#>  [81,] 0.73693543 0.26306457
#>  [82,] 0.72165234 0.27834766
#>  [83,] 0.90668838 0.09331162
#>  [84,] 0.82354059 0.17645941
#>  [85,] 0.82354059 0.17645941
#>  [86,] 0.73491252 0.26508748
#>  [87,] 0.80011658 0.19988342
#>  [88,] 0.69732759 0.30267241
#>  [89,] 0.81806030 0.18193970
#>  [90,] 0.90668838 0.09331162
#>  [91,] 0.70136438 0.29863562
#>  [92,] 0.67774936 0.32225064
#>  [93,] 0.77106098 0.22893902
#>  [94,] 0.79675782 0.20324218
#>  [95,] 1.00000000 0.00000000
#>  [96,] 0.89342820 0.10657180
#>  [97,] 0.78477459 0.21522541
#>  [98,] 0.79675782 0.20324218
#>  [99,] 0.78430641 0.21569359
#> [100,] 0.73491252 0.26508748
#> [101,] 0.72915554 0.27084446
#> [102,] 0.59112945 0.40887055
#> [103,] 0.90668838 0.09331162
#> [104,] 0.79803477 0.20196523
#> [105,] 0.79803477 0.20196523
#> [106,] 0.90668838 0.09331162
#> [107,] 0.79803477 0.20196523
#> [108,] 0.82220475 0.17779525
#> [109,] 0.78632933 0.21367067
#> [110,] 0.69301771 0.30698229
#> [111,] 0.59171098 0.40828902
#> [112,] 0.72889313 0.27110687
#> [113,] 0.71899171 0.28100829
#> [114,] 0.69524577 0.30475423
#> [115,] 0.70722900 0.29277100
#> [116,] 0.91137192 0.08862808
#> [117,] 0.81022152 0.18977848
#> [118,] 0.79373555 0.20626445
#> [119,] 0.89134639 0.10865361
#> [120,] 0.80389938 0.19610062
#> [121,] 0.91137192 0.08862808
#> [122,] 0.89134639 0.10865361
#> [123,] 1.00000000 0.00000000
#> [124,] 0.80819860 0.19180140
#> [125,] 0.80819860 0.19180140
#> [126,] 0.71488698 0.28511302
#> [127,] 0.83572734 0.16427266
#> [128,] 0.74241572 0.25758428
#> [129,] 0.91685221 0.08314779
#> [130,] 0.91685221 0.08314779
#> [131,] 0.74241572 0.25758428
#> [132,] 0.91685221 0.08314779
#> [133,] 0.91685221 0.08314779
#> [134,] 0.91685221 0.08314779
#> [135,] 0.91685221 0.08314779
#> [136,] 0.73905696 0.26094304
#> [137,] 0.82018183 0.17981817
#> [138,] 0.72277452 0.27722548
#> [139,] 0.80819860 0.19180140
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  4.7813906 14.3007655  8.2156545  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  2.7055800  0.0000000  2.9201012  0.7037383  4.3892676  1.0888118 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  1.5834121  4.1888910  2.1766230  0.0000000  0.0000000  0.0000000  2.5090593 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  4.8667179  3.8193392  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.7012702  3.3545065  1.1472353  0.0000000  0.0000000  2.0226678  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  0.0000000  7.9467233  0.0000000  6.3586850 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  2.1868823  0.0000000  3.4543915  0.0000000  6.6272405  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.6668741  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  1.9209973  1.3631744 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V12        V45        V51        V47        V31        V10 
#> 14.3007655  8.2156545  7.9467233  6.6272405  6.3586850  4.8667179  4.7813906 
#>        V20        V23        V32        V54         V5        V36        V19 
#>  4.3892676  4.1888910  3.8193392  3.6668741  3.4543915  3.3545065  2.9201012 
#>        V17        V28        V48        V24         V4         V8        V35 
#>  2.7055800  2.5090593  2.1868823  2.1766230  2.0226678  1.9209973  1.7012702 
#>        V22         V9        V37        V21         V2         V1        V13 
#>  1.5834121  1.3631744  1.1472353  1.0888118  0.7037383  0.0000000  0.0000000 
#>        V14        V15        V16        V18        V25        V26        V27 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V29         V3        V30        V33        V34        V38        V39 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V40        V41        V42        V43        V44        V46        V49 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V50        V52        V53        V55        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6        V60         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058