Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 65 M (0.53237410 0.46762590)  
#>   2) V11>=0.1707 86 21 M (0.75581395 0.24418605)  
#>     4) V16< 0.67795 69  9 M (0.86956522 0.13043478)  
#>       8) V48>=0.0451 59  2 M (0.96610169 0.03389831) *
#>       9) V48< 0.0451 10  3 R (0.30000000 0.70000000) *
#>     5) V16>=0.67795 17  5 R (0.29411765 0.70588235) *
#>   3) V11< 0.1707 53  9 R (0.16981132 0.83018868)  
#>     6) V47>=0.17365 9  3 M (0.66666667 0.33333333) *
#>     7) V47< 0.17365 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V11>=0.195 83 17 M (0.79518072 0.20481928)  
#>      4) V49>=0.02315 65  6 M (0.90769231 0.09230769) *
#>      5) V49< 0.02315 18  7 R (0.38888889 0.61111111) *
#>    3) V11< 0.195 56 15 R (0.26785714 0.73214286)  
#>      6) V44>=0.19395 21  8 M (0.61904762 0.38095238)  
#>       12) V33< 0.5303 12  0 M (1.00000000 0.00000000) *
#>       13) V33>=0.5303 9  1 R (0.11111111 0.88888889) *
#>      7) V44< 0.19395 35  2 R (0.05714286 0.94285714) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V51>=0.01235 77 21 M (0.72727273 0.27272727)  
#>      4) V12>=0.12565 61  8 M (0.86885246 0.13114754)  
#>        8) V42>=0.1019 53  3 M (0.94339623 0.05660377) *
#>        9) V42< 0.1019 8  3 R (0.37500000 0.62500000) *
#>      5) V12< 0.12565 16  3 R (0.18750000 0.81250000) *
#>    3) V51< 0.01235 62 15 R (0.24193548 0.75806452)  
#>      6) V22>=0.91995 9  1 M (0.88888889 0.11111111) *
#>      7) V22< 0.91995 53  7 R (0.13207547 0.86792453)  
#>       14) V28>=0.92415 7  2 M (0.71428571 0.28571429) *
#>       15) V28< 0.92415 46  2 R (0.04347826 0.95652174) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V21>=0.6946 67 18 M (0.73134328 0.26865672)  
#>      4) V32< 0.5232 54  9 M (0.83333333 0.16666667)  
#>        8) V44>=0.1384 39  1 M (0.97435897 0.02564103) *
#>        9) V44< 0.1384 15  7 R (0.46666667 0.53333333) *
#>      5) V32>=0.5232 13  4 R (0.30769231 0.69230769) *
#>    3) V21< 0.6946 72 19 R (0.26388889 0.73611111)  
#>      6) V36< 0.1596 11  0 M (1.00000000 0.00000000) *
#>      7) V36>=0.1596 61  8 R (0.13114754 0.86885246)  
#>       14) V47>=0.2697 7  2 M (0.71428571 0.28571429) *
#>       15) V47< 0.2697 54  3 R (0.05555556 0.94444444) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V37< 0.4731 104 29 M (0.72115385 0.27884615)  
#>      4) V58>=0.0062 62  4 M (0.93548387 0.06451613) *
#>      5) V58< 0.0062 42 17 R (0.40476190 0.59523810)  
#>       10) V52>=0.0131 13  1 M (0.92307692 0.07692308) *
#>       11) V52< 0.0131 29  5 R (0.17241379 0.82758621)  
#>         22) V49>=0.0424 9  4 M (0.55555556 0.44444444) *
#>         23) V49< 0.0424 20  0 R (0.00000000 1.00000000) *
#>    3) V37>=0.4731 35  6 R (0.17142857 0.82857143)  
#>      6) V35>=0.77565 7  2 M (0.71428571 0.28571429) *
#>      7) V35< 0.77565 28  1 R (0.03571429 0.96428571) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V31< 0.37765 57 10 M (0.82456140 0.17543860)  
#>      4) V30>=0.16725 50  4 M (0.92000000 0.08000000) *
#>      5) V30< 0.16725 7  1 R (0.14285714 0.85714286) *
#>    3) V31>=0.37765 82 31 R (0.37804878 0.62195122)  
#>      6) V45>=0.35835 11  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.35835 71 20 R (0.28169014 0.71830986)  
#>       14) V24>=0.69435 34 16 M (0.52941176 0.47058824)  
#>         28) V36< 0.5006 23  5 M (0.78260870 0.21739130)  
#>           56) V11>=0.17355 16  0 M (1.00000000 0.00000000) *
#>           57) V11< 0.17355 7  2 R (0.28571429 0.71428571) *
#>         29) V36>=0.5006 11  0 R (0.00000000 1.00000000) *
#>       15) V24< 0.69435 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V11>=0.2389 73 18 M (0.75342466 0.24657534)  
#>      4) V32>=0.08755 66 11 M (0.83333333 0.16666667)  
#>        8) V51>=0.01235 42  1 M (0.97619048 0.02380952) *
#>        9) V51< 0.01235 24 10 M (0.58333333 0.41666667)  
#>         18) V16< 0.2341 11  0 M (1.00000000 0.00000000) *
#>         19) V16>=0.2341 13  3 R (0.23076923 0.76923077) *
#>      5) V32< 0.08755 7  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.2389 66 21 R (0.31818182 0.68181818)  
#>      6) V14>=0.2798 24  7 M (0.70833333 0.29166667)  
#>       12) V14< 0.3177 11  0 M (1.00000000 0.00000000) *
#>       13) V14>=0.3177 13  6 R (0.46153846 0.53846154) *
#>      7) V14< 0.2798 42  4 R (0.09523810 0.90476190) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V21>=0.56525 96 30 M (0.68750000 0.31250000)  
#>      4) V6< 0.1695 83 19 M (0.77108434 0.22891566)  
#>        8) V38>=0.14795 73 11 M (0.84931507 0.15068493)  
#>         16) V31>=0.2292 64  5 M (0.92187500 0.07812500) *
#>         17) V31< 0.2292 9  3 R (0.33333333 0.66666667) *
#>        9) V38< 0.14795 10  2 R (0.20000000 0.80000000) *
#>      5) V6>=0.1695 13  2 R (0.15384615 0.84615385) *
#>    3) V21< 0.56525 43 11 R (0.25581395 0.74418605)  
#>      6) V15< 0.13635 8  0 M (1.00000000 0.00000000) *
#>      7) V15>=0.13635 35  3 R (0.08571429 0.91428571) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 48 M (0.65467626 0.34532374)  
#>    2) V11>=0.1089 124 33 M (0.73387097 0.26612903)  
#>      4) V52>=0.0093 85  9 M (0.89411765 0.10588235)  
#>        8) V50< 0.04465 77  4 M (0.94805195 0.05194805) *
#>        9) V50>=0.04465 8  3 R (0.37500000 0.62500000) *
#>      5) V52< 0.0093 39 15 R (0.38461538 0.61538462)  
#>       10) V49>=0.0596 13  2 M (0.84615385 0.15384615) *
#>       11) V49< 0.0596 26  4 R (0.15384615 0.84615385) *
#>    3) V11< 0.1089 15  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 47 M (0.66187050 0.33812950)  
#>   2) V37< 0.48 105 22 M (0.79047619 0.20952381)  
#>     4) V12>=0.1558 87  8 M (0.90804598 0.09195402) *
#>     5) V12< 0.1558 18  4 R (0.22222222 0.77777778) *
#>   3) V37>=0.48 34  9 R (0.26470588 0.73529412)  
#>     6) V44>=0.35425 8  0 M (1.00000000 0.00000000) *
#>     7) V44< 0.35425 26  1 R (0.03846154 0.96153846) *
#> 
#> 
#> $weights
#>  [1] 0.7586613 0.7056483 0.6763787 0.5780631 0.7266919 0.8524045 0.8134833
#>  [8] 0.3487368 0.7076599 0.6106338
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.1297654 4.6485962
#>   [2,] 1.2857230 5.4926386
#>   [3,] 2.2874445 4.4909171
#>   [4,] 2.1004958 4.6778659
#>   [5,] 1.4663212 5.3120404
#>   [6,] 1.8943452 4.8840164
#>   [7,] 2.3937224 4.3846392
#>   [8,] 0.5780631 6.2002986
#>   [9,] 0.3487368 6.4296249
#>  [10,] 1.0754287 5.7029329
#>  [11,] 1.5580528 5.2203088
#>  [12,] 2.1617320 4.6166297
#>  [13,] 1.3840386 5.3943230
#>  [14,] 2.8953786 3.8829831
#>  [15,] 1.4133083 5.3650534
#>  [16,] 2.8655076 3.9128541
#>  [17,] 0.8524045 5.9259571
#>  [18,] 1.3182937 5.4600679
#>  [19,] 1.8130464 4.9653152
#>  [20,] 2.7677933 4.0105683
#>  [21,] 2.2867564 4.4916053
#>  [22,] 1.0754287 5.7029329
#>  [23,] 3.0680895 3.7102721
#>  [24,] 1.1073981 5.6709636
#>  [25,] 2.8973902 3.8809715
#>  [26,] 1.7830886 4.9952730
#>  [27,] 1.0754287 5.7029329
#>  [28,] 1.4133083 5.3650534
#>  [29,] 0.7076599 6.0707017
#>  [30,] 0.7266919 6.0516697
#>  [31,] 1.3162821 5.4620795
#>  [32,] 1.4133083 5.3650534
#>  [33,] 0.9593705 5.8189911
#>  [34,] 2.4805138 4.2978479
#>  [35,] 2.0423727 4.7359889
#>  [36,] 1.5401752 5.2381864
#>  [37,] 0.0000000 6.7783616
#>  [38,] 1.6860625 5.0922991
#>  [39,] 0.0000000 6.7783616
#>  [40,] 1.4030706 5.3752910
#>  [41,] 2.8980783 3.8802833
#>  [42,] 0.3487368 6.4296249
#>  [43,] 0.3487368 6.4296249
#>  [44,] 0.3487368 6.4296249
#>  [45,] 1.0754287 5.7029329
#>  [46,] 0.7266919 6.0516697
#>  [47,] 1.0754287 5.7029329
#>  [48,] 0.8524045 5.9259571
#>  [49,] 0.8524045 5.9259571
#>  [50,] 1.5790965 5.1992652
#>  [51,] 1.1622200 5.6161416
#>  [52,] 1.7402831 5.0380785
#>  [53,] 1.4030706 5.3752910
#>  [54,] 1.5790965 5.1992652
#>  [55,] 0.8134833 5.9648783
#>  [56,] 2.7506340 4.0277276
#>  [57,] 3.0993708 3.6789909
#>  [58,] 1.3162821 5.4620795
#>  [59,] 1.4133083 5.3650534
#>  [60,] 2.2657128 4.5126489
#>  [61,] 2.2267916 4.5515701
#>  [62,] 1.4343519 5.3440098
#>  [63,] 2.0449857 4.7333760
#>  [64,] 1.9153888 4.8629728
#>  [65,] 0.3487368 6.4296249
#>  [66,] 1.0754287 5.7029329
#>  [67,] 5.8515618 0.9267998
#>  [68,] 5.3433216 1.4350400
#>  [69,] 4.7652586 2.0131031
#>  [70,] 4.0283290 2.7500326
#>  [71,] 4.0047882 2.7735734
#>  [72,] 5.3120404 1.4663212
#>  [73,] 4.5298383 2.2485233
#>  [74,] 4.6356617 2.1426999
#>  [75,] 3.8254613 2.9529003
#>  [76,] 5.6709636 1.1073981
#>  [77,] 4.1518319 2.6265297
#>  [78,] 4.8574803 1.9208814
#>  [79,] 6.4296249 0.3487368
#>  [80,] 4.2028333 2.5755283
#>  [81,] 5.6161416 1.1622200
#>  [82,] 5.6161416 1.1622200
#>  [83,] 6.2002986 0.5780631
#>  [84,] 5.3440098 1.4343519
#>  [85,] 5.9648783 0.8134833
#>  [86,] 6.0707017 0.7076599
#>  [87,] 6.2002986 0.5780631
#>  [88,] 5.6709636 1.1073981
#>  [89,] 5.4090665 1.3692951
#>  [90,] 5.6709636 1.1073981
#>  [91,] 3.5328361 3.2455256
#>  [92,] 4.7836892 1.9946724
#>  [93,] 5.4410359 1.3373257
#>  [94,] 6.0707017 0.7076599
#>  [95,] 6.1677278 0.6106338
#>  [96,] 5.1426124 1.6357492
#>  [97,] 6.2002986 0.5780631
#>  [98,] 5.6161416 1.1622200
#>  [99,] 5.7219650 1.0563967
#> [100,] 6.2002986 0.5780631
#> [101,] 6.4296249 0.3487368
#> [102,] 3.7293041 3.0490575
#> [103,] 5.7219650 1.0563967
#> [104,] 5.9259571 0.8524045
#> [105,] 4.3324302 2.4459315
#> [106,] 4.4360951 2.3422665
#> [107,] 4.3546814 2.4236802
#> [108,] 5.2381864 1.5401752
#> [109,] 5.4736066 1.3047550
#> [110,] 5.1248699 1.6534918
#> [111,] 5.4736066 1.3047550
#> [112,] 5.2381864 1.5401752
#> [113,] 5.0380785 1.7402831
#> [114,] 5.5239199 1.2544418
#> [115,] 6.2002986 0.5780631
#> [116,] 5.3542446 1.4241171
#> [117,] 4.5618077 2.2165539
#> [118,] 4.4165218 2.3618398
#> [119,] 6.1019829 0.6763787
#> [120,] 4.5853485 2.1930132
#> [121,] 4.0580887 2.7202729
#> [122,] 5.1751831 1.6031785
#> [123,] 6.4296249 0.3487368
#> [124,] 6.2002986 0.5780631
#> [125,] 6.2002986 0.5780631
#> [126,] 6.2002986 0.5780631
#> [127,] 4.9397629 1.8385988
#> [128,] 5.5772204 1.2011413
#> [129,] 4.8505284 1.9278332
#> [130,] 6.0707017 0.7076599
#> [131,] 6.4296249 0.3487368
#> [132,] 5.9648783 0.8134833
#> [133,] 6.4296249 0.3487368
#> [134,] 5.7532462 1.0251155
#> [135,] 6.4296249 0.3487368
#> [136,] 5.7532462 1.0251155
#> [137,] 6.7783616 0.0000000
#> [138,] 6.1019829 0.6763787
#> [139,] 4.7359889 2.0423727
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.31420062 0.68579938
#>   [2,] 0.18968049 0.81031951
#>   [3,] 0.33746274 0.66253726
#>   [4,] 0.30988252 0.69011748
#>   [5,] 0.21632384 0.78367616
#>   [6,] 0.27946948 0.72053052
#>   [7,] 0.35314174 0.64685826
#>   [8,] 0.08528065 0.91471935
#>   [9,] 0.05144853 0.94855147
#>  [10,] 0.15865614 0.84134386
#>  [11,] 0.22985685 0.77014315
#>  [12,] 0.31891659 0.68108341
#>  [13,] 0.20418483 0.79581517
#>  [14,] 0.42715021 0.57284979
#>  [15,] 0.20850293 0.79149707
#>  [16,] 0.42274339 0.57725661
#>  [17,] 0.12575377 0.87424623
#>  [18,] 0.19448560 0.80551440
#>  [19,] 0.26747561 0.73252439
#>  [20,] 0.40832777 0.59167223
#>  [21,] 0.33736122 0.66263878
#>  [22,] 0.15865614 0.84134386
#>  [23,] 0.45262996 0.54737004
#>  [24,] 0.16337253 0.83662747
#>  [25,] 0.42744697 0.57255303
#>  [26,] 0.26305599 0.73694401
#>  [27,] 0.15865614 0.84134386
#>  [28,] 0.20850293 0.79149707
#>  [29,] 0.10439985 0.89560015
#>  [30,] 0.10720761 0.89279239
#>  [31,] 0.19418883 0.80581117
#>  [32,] 0.20850293 0.79149707
#>  [33,] 0.14153428 0.85846572
#>  [34,] 0.36594592 0.63405408
#>  [35,] 0.30130772 0.69869228
#>  [36,] 0.22721940 0.77278060
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.24874189 0.75125811
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.20699259 0.79300741
#>  [41,] 0.42754849 0.57245151
#>  [42,] 0.05144853 0.94855147
#>  [43,] 0.05144853 0.94855147
#>  [44,] 0.05144853 0.94855147
#>  [45,] 0.15865614 0.84134386
#>  [46,] 0.10720761 0.89279239
#>  [47,] 0.15865614 0.84134386
#>  [48,] 0.12575377 0.87424623
#>  [49,] 0.12575377 0.87424623
#>  [50,] 0.23296138 0.76703862
#>  [51,] 0.17146032 0.82853968
#>  [52,] 0.25674097 0.74325903
#>  [53,] 0.20699259 0.79300741
#>  [54,] 0.23296138 0.76703862
#>  [55,] 0.12001179 0.87998821
#>  [56,] 0.40579629 0.59420371
#>  [57,] 0.45724482 0.54275518
#>  [58,] 0.19418883 0.80581117
#>  [59,] 0.20850293 0.79149707
#>  [60,] 0.33425670 0.66574330
#>  [61,] 0.32851472 0.67148528
#>  [62,] 0.21160746 0.78839254
#>  [63,] 0.30169321 0.69830679
#>  [64,] 0.28257401 0.71742599
#>  [65,] 0.05144853 0.94855147
#>  [66,] 0.15865614 0.84134386
#>  [67,] 0.86327082 0.13672918
#>  [68,] 0.78829102 0.21170898
#>  [69,] 0.70301038 0.29698962
#>  [70,] 0.59429243 0.40570757
#>  [71,] 0.59081950 0.40918050
#>  [72,] 0.78367616 0.21632384
#>  [73,] 0.66827924 0.33172076
#>  [74,] 0.68389117 0.31610883
#>  [75,] 0.56436371 0.43563629
#>  [76,] 0.83662747 0.16337253
#>  [77,] 0.61251260 0.38748740
#>  [78,] 0.71661569 0.28338431
#>  [79,] 0.94855147 0.05144853
#>  [80,] 0.62003675 0.37996325
#>  [81,] 0.82853968 0.17146032
#>  [82,] 0.82853968 0.17146032
#>  [83,] 0.91471935 0.08528065
#>  [84,] 0.78839254 0.21160746
#>  [85,] 0.87998821 0.12001179
#>  [86,] 0.89560015 0.10439985
#>  [87,] 0.91471935 0.08528065
#>  [88,] 0.83662747 0.16337253
#>  [89,] 0.79799026 0.20200974
#>  [90,] 0.83662747 0.16337253
#>  [91,] 0.52119321 0.47880679
#>  [92,] 0.70572942 0.29427058
#>  [93,] 0.80270664 0.19729336
#>  [94,] 0.89560015 0.10439985
#>  [95,] 0.90991425 0.09008575
#>  [96,] 0.75868073 0.24131927
#>  [97,] 0.91471935 0.08528065
#>  [98,] 0.82853968 0.17146032
#>  [99,] 0.84415162 0.15584838
#> [100,] 0.91471935 0.08528065
#> [101,] 0.94855147 0.05144853
#> [102,] 0.55017780 0.44982220
#> [103,] 0.84415162 0.15584838
#> [104,] 0.87424623 0.12575377
#> [105,] 0.63915595 0.36084405
#> [106,] 0.65444946 0.34555054
#> [107,] 0.64243864 0.35756136
#> [108,] 0.77278060 0.22721940
#> [109,] 0.80751174 0.19248826
#> [110,] 0.75606321 0.24393679
#> [111,] 0.80751174 0.19248826
#> [112,] 0.77278060 0.22721940
#> [113,] 0.74325903 0.25674097
#> [114,] 0.81493437 0.18506563
#> [115,] 0.91471935 0.08528065
#> [116,] 0.78990246 0.21009754
#> [117,] 0.67299562 0.32700438
#> [118,] 0.65156184 0.34843816
#> [119,] 0.90021502 0.09978498
#> [120,] 0.67646855 0.32353145
#> [121,] 0.59868283 0.40131717
#> [122,] 0.76348584 0.23651416
#> [123,] 0.94855147 0.05144853
#> [124,] 0.91471935 0.08528065
#> [125,] 0.91471935 0.08528065
#> [126,] 0.91471935 0.08528065
#> [127,] 0.72875470 0.27124530
#> [128,] 0.82279770 0.17720230
#> [129,] 0.71559009 0.28440991
#> [130,] 0.89560015 0.10439985
#> [131,] 0.94855147 0.05144853
#> [132,] 0.87998821 0.12001179
#> [133,] 0.94855147 0.05144853
#> [134,] 0.84876648 0.15123352
#> [135,] 0.94855147 0.05144853
#> [136,] 0.84876648 0.15123352
#> [137,] 1.00000000 0.00000000
#> [138,] 0.90021502 0.09978498
#> [139,] 0.69869228 0.30130772
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  0.0000000 17.8069537  5.3193245  0.0000000  3.9107228  1.2218954 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  4.0518380  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  4.0641615 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  1.9185677  0.0000000  2.1968074  0.0000000  0.0000000  0.0000000  1.1902784 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  2.0348084  4.2920219  3.3785853  1.8457404  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.2061807  5.1185106  6.4924600  0.8324082  0.0000000  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.9777125  0.0000000  5.1445595  2.6964318  0.0000000  2.3074840 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.8527414  4.5116671  0.0000000  1.0840698  4.7566011  5.5272736  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  0.0000000  3.2989443  0.0000000  0.9612500 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 73 66 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V37        V52        V12        V44        V36        V51 
#> 17.8069537  6.4924600  5.5272736  5.3193245  5.1445595  5.1185106  4.7566011 
#>        V49        V31        V21        V16        V14        V32        V58 
#>  4.5116671  4.2920219  4.0641615  4.0518380  3.9107228  3.3785853  3.2989443 
#>        V45        V47        V24        V30        V22        V48        V33 
#>  2.6964318  2.3074840  2.1968074  2.0348084  1.9185677  1.8527414  1.8457404 
#>        V15        V35        V28        V50        V42         V6        V38 
#>  1.2218954  1.2061807  1.1902784  1.0840698  0.9777125  0.9612500  0.8324082 
#>         V1        V10        V13        V17        V18        V19         V2 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V20        V23        V25        V26        V27        V29         V3 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V34        V39         V4        V40        V41        V43        V46 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V5        V53        V54        V55        V56        V57        V59 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986