Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.4676259 0.5323741)  
#>    2) V12>=0.22505 78 22 M (0.7179487 0.2820513)  
#>      4) V54>=0.00815 51  6 M (0.8823529 0.1176471) *
#>      5) V54< 0.00815 27 11 R (0.4074074 0.5925926)  
#>       10) V34< 0.2197 9  0 M (1.0000000 0.0000000) *
#>       11) V34>=0.2197 18  2 R (0.1111111 0.8888889) *
#>    3) V12< 0.22505 61  9 R (0.1475410 0.8524590)  
#>      6) V21>=0.80055 14  6 M (0.5714286 0.4285714) *
#>      7) V21< 0.80055 47  1 R (0.0212766 0.9787234) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.4748201 0.5251799)  
#>    2) V9>=0.10375 99 33 M (0.6666667 0.3333333)  
#>      4) V48>=0.076 44  0 M (1.0000000 0.0000000) *
#>      5) V48< 0.076 55 22 R (0.4000000 0.6000000)  
#>       10) V9>=0.18095 18  3 M (0.8333333 0.1666667) *
#>       11) V9< 0.18095 37  7 R (0.1891892 0.8108108)  
#>         22) V38>=0.4152 11  4 M (0.6363636 0.3636364) *
#>         23) V38< 0.4152 26  0 R (0.0000000 1.0000000) *
#>    3) V9< 0.10375 40  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V12>=0.16495 98 35 M (0.64285714 0.35714286)  
#>      4) V33< 0.4975 65 12 M (0.81538462 0.18461538)  
#>        8) V51>=0.01235 41  0 M (1.00000000 0.00000000) *
#>        9) V51< 0.01235 24 12 M (0.50000000 0.50000000)  
#>         18) V43>=0.2164 9  0 M (1.00000000 0.00000000) *
#>         19) V43< 0.2164 15  3 R (0.20000000 0.80000000) *
#>      5) V33>=0.4975 33 10 R (0.30303030 0.69696970)  
#>       10) V49>=0.07215 8  0 M (1.00000000 0.00000000) *
#>       11) V49< 0.07215 25  2 R (0.08000000 0.92000000) *
#>    3) V12< 0.16495 41  3 R (0.07317073 0.92682927) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V45>=0.26515 35  1 M (0.97142857 0.02857143) *
#>    3) V45< 0.26515 104 43 R (0.41346154 0.58653846)  
#>      6) V12>=0.1786 71 30 M (0.57746479 0.42253521)  
#>       12) V40< 0.37815 55 15 M (0.72727273 0.27272727)  
#>         24) V54>=0.008 29  1 M (0.96551724 0.03448276) *
#>         25) V54< 0.008 26 12 R (0.46153846 0.53846154)  
#>           50) V15< 0.2994 15  3 M (0.80000000 0.20000000) *
#>           51) V15>=0.2994 11  0 R (0.00000000 1.00000000) *
#>       13) V40>=0.37815 16  1 R (0.06250000 0.93750000) *
#>      7) V12< 0.1786 33  2 R (0.06060606 0.93939394) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V49>=0.04605 64 12 M (0.81250000 0.18750000)  
#>      4) V9>=0.09405 57  5 M (0.91228070 0.08771930)  
#>        8) V17< 0.54945 50  1 M (0.98000000 0.02000000) *
#>        9) V17>=0.54945 7  3 R (0.42857143 0.57142857) *
#>      5) V9< 0.09405 7  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.04605 75 29 R (0.38666667 0.61333333)  
#>      6) V21>=0.84825 18  3 M (0.83333333 0.16666667) *
#>      7) V21< 0.84825 57 14 R (0.24561404 0.75438596)  
#>       14) V26>=0.99215 7  0 M (1.00000000 0.00000000) *
#>       15) V26< 0.99215 50  7 R (0.14000000 0.86000000)  
#>         30) V42>=0.3041 11  5 M (0.54545455 0.45454545) *
#>         31) V42< 0.3041 39  1 R (0.02564103 0.97435897) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 R (0.48920863 0.51079137)  
#>   2) V20>=0.57355 80 26 M (0.67500000 0.32500000)  
#>     4) V42>=0.08525 63  9 M (0.85714286 0.14285714)  
#>       8) V56< 0.01955 56  4 M (0.92857143 0.07142857) *
#>       9) V56>=0.01955 7  2 R (0.28571429 0.71428571) *
#>     5) V42< 0.08525 17  0 R (0.00000000 1.00000000) *
#>   3) V20< 0.57355 59 14 R (0.23728814 0.76271186)  
#>     6) V26>=0.9885 8  1 M (0.87500000 0.12500000) *
#>     7) V26< 0.9885 51  7 R (0.13725490 0.86274510) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 M (0.52517986 0.47482014)  
#>   2) V9>=0.16605 62 13 M (0.79032258 0.20967742)  
#>     4) V15< 0.5312 49  4 M (0.91836735 0.08163265) *
#>     5) V15>=0.5312 13  4 R (0.30769231 0.69230769) *
#>   3) V9< 0.16605 77 24 R (0.31168831 0.68831169)  
#>     6) V59>=0.0099 19  4 M (0.78947368 0.21052632) *
#>     7) V59< 0.0099 58  9 R (0.15517241 0.84482759) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V36< 0.50725 91 22 M (0.75824176 0.24175824)  
#>      4) V49>=0.036 61  5 M (0.91803279 0.08196721)  
#>        8) V16< 0.7732 54  0 M (1.00000000 0.00000000) *
#>        9) V16>=0.7732 7  2 R (0.28571429 0.71428571) *
#>      5) V49< 0.036 30 13 R (0.43333333 0.56666667)  
#>       10) V1>=0.02845 8  0 M (1.00000000 0.00000000) *
#>       11) V1< 0.02845 22  5 R (0.22727273 0.77272727)  
#>         22) V25< 0.5805 9  4 M (0.55555556 0.44444444) *
#>         23) V25>=0.5805 13  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.50725 48 13 R (0.27083333 0.72916667)  
#>      6) V60>=0.01325 10  0 M (1.00000000 0.00000000) *
#>      7) V60< 0.01325 38  3 R (0.07894737 0.92105263) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 51 M (0.63309353 0.36690647)  
#>   2) V37< 0.47795 93 17 M (0.81720430 0.18279570)  
#>     4) V33>=0.1432 86 11 M (0.87209302 0.12790698)  
#>       8) V12>=0.14715 76  5 M (0.93421053 0.06578947) *
#>       9) V12< 0.14715 10  4 R (0.40000000 0.60000000) *
#>     5) V33< 0.1432 7  1 R (0.14285714 0.85714286) *
#>   3) V37>=0.47795 46 12 R (0.26086957 0.73913043)  
#>     6) V1>=0.0482 9  0 M (1.00000000 0.00000000) *
#>     7) V1< 0.0482 37  3 R (0.08108108 0.91891892) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 56 M (0.59712230 0.40287770)  
#>   2) V13>=0.23095 96 21 M (0.78125000 0.21875000)  
#>     4) V5>=0.03405 88 13 M (0.85227273 0.14772727) *
#>     5) V5< 0.03405 8  0 R (0.00000000 1.00000000) *
#>   3) V13< 0.23095 43  8 R (0.18604651 0.81395349)  
#>     6) V45>=0.19425 9  3 M (0.66666667 0.33333333) *
#>     7) V45< 0.19425 34  2 R (0.05882353 0.94117647) *
#> 
#> 
#> $weights
#>  [1] 0.8916956 0.7533849 1.0589868 1.0974062 0.7685126 0.5437766 0.6762952
#>  [8] 0.9332104 0.7177342 0.6169122
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0465924 6.0113226
#>   [2,] 3.1470182 4.9108967
#>   [3,] 3.1952032 4.8627117
#>   [4,] 2.6475289 5.4103860
#>   [5,] 3.0212418 5.0366731
#>   [6,] 1.5218976 6.5360173
#>   [7,] 2.0938993 5.9640156
#>   [8,] 0.0000000 8.0579149
#>   [9,] 0.0000000 8.0579149
#>  [10,] 1.2971616 6.7607533
#>  [11,] 1.6602083 6.3977067
#>  [12,] 2.5903690 5.4675459
#>  [13,] 2.1388098 5.9191051
#>  [14,] 1.2932074 6.7647075
#>  [15,] 1.9734568 6.0844582
#>  [16,] 2.7643265 5.2935884
#>  [17,] 0.6169122 7.4410027
#>  [18,] 0.0000000 8.0579149
#>  [19,] 0.5437766 7.5141383
#>  [20,] 2.3936332 5.6642817
#>  [21,] 2.6653970 5.3925179
#>  [22,] 0.0000000 8.0579149
#>  [23,] 2.9217191 5.1361958
#>  [24,] 2.6684925 5.3894225
#>  [25,] 0.0000000 8.0579149
#>  [26,] 0.0000000 8.0579149
#>  [27,] 0.0000000 8.0579149
#>  [28,] 1.4296801 6.6282348
#>  [29,] 0.0000000 8.0579149
#>  [30,] 0.0000000 8.0579149
#>  [31,] 1.5086078 6.5493071
#>  [32,] 2.2039849 5.8539300
#>  [33,] 1.8659189 6.1919961
#>  [34,] 2.3779424 5.6799725
#>  [35,] 1.7143184 6.3435965
#>  [36,] 0.0000000 8.0579149
#>  [37,] 2.4320526 5.6258623
#>  [38,] 0.9332104 7.1247045
#>  [39,] 0.7177342 7.3401807
#>  [40,] 1.7017231 6.3561918
#>  [41,] 2.7483508 5.3095641
#>  [42,] 1.3122893 6.7456256
#>  [43,] 0.0000000 8.0579149
#>  [44,] 1.9885845 6.0693304
#>  [45,] 0.0000000 8.0579149
#>  [46,] 0.0000000 8.0579149
#>  [47,] 0.7685126 7.2894023
#>  [48,] 0.0000000 8.0579149
#>  [49,] 0.8916956 7.1662193
#>  [50,] 1.3122893 6.7456256
#>  [51,] 0.0000000 8.0579149
#>  [52,] 1.3122893 6.7456256
#>  [53,] 0.8916956 7.1662193
#>  [54,] 1.6602083 6.3977067
#>  [55,] 1.6602083 6.3977067
#>  [56,] 2.6318080 5.4261069
#>  [57,] 2.6318080 5.4261069
#>  [58,] 1.3702972 6.6876178
#>  [59,] 1.3702972 6.6876178
#>  [60,] 0.7533849 7.3045300
#>  [61,] 1.3702972 6.6876178
#>  [62,] 1.3702972 6.6876178
#>  [63,] 1.8151404 6.2427745
#>  [64,] 1.3346464 6.7232685
#>  [65,] 2.3628906 5.6950244
#>  [66,] 1.6865954 6.3713195
#>  [67,] 6.4167321 1.6411828
#>  [68,] 7.5141383 0.5437766
#>  [69,] 6.6131071 1.4448078
#>  [70,] 6.6131071 1.4448078
#>  [71,] 6.7456256 1.3122893
#>  [72,] 5.8597221 2.1981928
#>  [73,] 6.9989281 1.0589868
#>  [74,] 5.8275429 2.2303720
#>  [75,] 7.4410027 0.6169122
#>  [76,] 5.2739133 2.7840016
#>  [77,] 6.5360173 1.5218976
#>  [78,] 6.3713195 1.6865954
#>  [79,] 6.5716681 1.4862468
#>  [80,] 5.8953729 2.1625420
#>  [81,] 6.2071238 1.8507911
#>  [82,] 7.2894023 0.7685126
#>  [83,] 6.6282348 1.4296801
#>  [84,] 6.6131071 1.4448078
#>  [85,] 7.2894023 0.7685126
#>  [86,] 7.2894023 0.7685126
#>  [87,] 6.3561918 1.7017231
#>  [88,] 5.6798966 2.3780183
#>  [89,] 6.0657176 1.9921973
#>  [90,] 6.5809278 1.4769871
#>  [91,] 6.5809278 1.4769871
#>  [92,] 5.5541202 2.5037947
#>  [93,] 5.7374173 2.3204977
#>  [94,] 6.9989281 1.0589868
#>  [95,] 5.7998198 2.2580951
#>  [96,] 5.9922407 2.0656742
#>  [97,] 6.8378431 1.2200718
#>  [98,] 7.5141383 0.5437766
#>  [99,] 5.8382392 2.2196757
#> [100,] 6.4551514 1.6027635
#> [101,] 4.8457216 3.2121933
#> [102,] 7.5141383 0.5437766
#> [103,] 6.2427745 1.8151404
#> [104,] 6.8378431 1.2200718
#> [105,] 5.2846096 2.7733053
#> [106,] 5.1044992 2.9534157
#> [107,] 5.1837877 2.8741272
#> [108,] 5.8631936 2.1947213
#> [109,] 5.8631936 2.1947213
#> [110,] 4.9714980 3.0864169
#> [111,] 5.9461475 2.1117674
#> [112,] 7.1662193 0.8916956
#> [113,] 5.7306751 2.3272398
#> [114,] 6.9989281 1.0589868
#> [115,] 5.9015219 2.1563930
#> [116,] 7.5141383 0.5437766
#> [117,] 7.1662193 0.8916956
#> [118,] 6.6724900 1.3854249
#> [119,] 7.1247045 0.9332104
#> [120,] 7.5141383 0.5437766
#> [121,] 8.0579149 0.0000000
#> [122,] 8.0579149 0.0000000
#> [123,] 8.0579149 0.0000000
#> [124,] 7.5141383 0.5437766
#> [125,] 6.6224427 1.4354723
#> [126,] 6.8378431 1.2200718
#> [127,] 5.9461475 2.1117674
#> [128,] 6.6224427 1.4354723
#> [129,] 8.0579149 0.0000000
#> [130,] 8.0579149 0.0000000
#> [131,] 7.3816197 0.6762952
#> [132,] 7.3816197 0.6762952
#> [133,] 8.0579149 0.0000000
#> [134,] 8.0579149 0.0000000
#> [135,] 7.5141383 0.5437766
#> [136,] 5.5692480 2.4886669
#> [137,] 5.7057207 2.3521943
#> [138,] 6.3820159 1.6758991
#> [139,] 6.5077922 1.5501227
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.25398535 0.74601465
#>   [2,] 0.39054994 0.60945006
#>   [3,] 0.39652978 0.60347022
#>   [4,] 0.32856253 0.67143747
#>   [5,] 0.37494089 0.62505911
#>   [6,] 0.18886990 0.81113010
#>   [7,] 0.25985622 0.74014378
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.00000000 1.00000000
#>  [10,] 0.16097980 0.83902020
#>  [11,] 0.20603447 0.79396553
#>  [12,] 0.32146889 0.67853111
#>  [13,] 0.26542968 0.73457032
#>  [14,] 0.16048909 0.83951091
#>  [15,] 0.24490911 0.75509089
#>  [16,] 0.34305730 0.65694270
#>  [17,] 0.07655978 0.92344022
#>  [18,] 0.00000000 1.00000000
#>  [19,] 0.06748354 0.93251646
#>  [20,] 0.29705368 0.70294632
#>  [21,] 0.33077999 0.66922001
#>  [22,] 0.00000000 1.00000000
#>  [23,] 0.36258997 0.63741003
#>  [24,] 0.33116414 0.66883586
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.17742557 0.82257443
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.00000000 1.00000000
#>  [31,] 0.18722062 0.81277938
#>  [32,] 0.27351802 0.72648198
#>  [33,] 0.23156348 0.76843652
#>  [34,] 0.29510642 0.70489358
#>  [35,] 0.21274963 0.78725037
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.30182158 0.69817842
#>  [38,] 0.11581289 0.88418711
#>  [39,] 0.08907195 0.91092805
#>  [40,] 0.21118653 0.78881347
#>  [41,] 0.34107469 0.65892531
#>  [42,] 0.16285718 0.83714282
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.24678648 0.75321352
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.09537364 0.90462636
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.11066084 0.88933916
#>  [50,] 0.16285718 0.83714282
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.16285718 0.83714282
#>  [53,] 0.11066084 0.88933916
#>  [54,] 0.20603447 0.79396553
#>  [55,] 0.20603447 0.79396553
#>  [56,] 0.32661154 0.67338846
#>  [57,] 0.32661154 0.67338846
#>  [58,] 0.17005605 0.82994395
#>  [59,] 0.17005605 0.82994395
#>  [60,] 0.09349626 0.90650374
#>  [61,] 0.17005605 0.82994395
#>  [62,] 0.17005605 0.82994395
#>  [63,] 0.22526180 0.77473820
#>  [64,] 0.16563173 0.83436827
#>  [65,] 0.29323846 0.70676154
#>  [66,] 0.20930916 0.79069084
#>  [67,] 0.79632661 0.20367339
#>  [68,] 0.93251646 0.06748354
#>  [69,] 0.82069706 0.17930294
#>  [70,] 0.82069706 0.17930294
#>  [71,] 0.83714282 0.16285718
#>  [72,] 0.72720080 0.27279920
#>  [73,] 0.86857806 0.13142194
#>  [74,] 0.72320730 0.27679270
#>  [75,] 0.92344022 0.07655978
#>  [76,] 0.65450100 0.34549900
#>  [77,] 0.81113010 0.18886990
#>  [78,] 0.79069084 0.20930916
#>  [79,] 0.81555441 0.18444559
#>  [80,] 0.73162511 0.26837489
#>  [81,] 0.77031389 0.22968611
#>  [82,] 0.90462636 0.09537364
#>  [83,] 0.82257443 0.17742557
#>  [84,] 0.82069706 0.17930294
#>  [85,] 0.90462636 0.09537364
#>  [86,] 0.90462636 0.09537364
#>  [87,] 0.78881347 0.21118653
#>  [88,] 0.70488417 0.29511583
#>  [89,] 0.75276516 0.24723484
#>  [90,] 0.81670356 0.18329644
#>  [91,] 0.81670356 0.18329644
#>  [92,] 0.68927512 0.31072488
#>  [93,] 0.71202257 0.28797743
#>  [94,] 0.86857806 0.13142194
#>  [95,] 0.71976683 0.28023317
#>  [96,] 0.74364656 0.25635344
#>  [97,] 0.84858715 0.15141285
#>  [98,] 0.93251646 0.06748354
#>  [99,] 0.72453473 0.27546527
#> [100,] 0.80109452 0.19890548
#> [101,] 0.60136173 0.39863827
#> [102,] 0.93251646 0.06748354
#> [103,] 0.77473820 0.22526180
#> [104,] 0.84858715 0.15141285
#> [105,] 0.65582843 0.34417157
#> [106,] 0.63347644 0.36652356
#> [107,] 0.64331626 0.35668374
#> [108,] 0.72763162 0.27236838
#> [109,] 0.72763162 0.27236838
#> [110,] 0.61697078 0.38302922
#> [111,] 0.73792632 0.26207368
#> [112,] 0.88933916 0.11066084
#> [113,] 0.71118585 0.28881415
#> [114,] 0.86857806 0.13142194
#> [115,] 0.73238821 0.26761179
#> [116,] 0.93251646 0.06748354
#> [117,] 0.88933916 0.11066084
#> [118,] 0.82806658 0.17193342
#> [119,] 0.88418711 0.11581289
#> [120,] 0.93251646 0.06748354
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 1.00000000 0.00000000
#> [124,] 0.93251646 0.06748354
#> [125,] 0.82185562 0.17814438
#> [126,] 0.84858715 0.15141285
#> [127,] 0.73792632 0.26207368
#> [128,] 0.82185562 0.17814438
#> [129,] 1.00000000 0.00000000
#> [130,] 1.00000000 0.00000000
#> [131,] 0.91607070 0.08392930
#> [132,] 0.91607070 0.08392930
#> [133,] 1.00000000 0.00000000
#> [134,] 1.00000000 0.00000000
#> [135,] 0.93251646 0.06748354
#> [136,] 0.69115249 0.30884751
#> [137,] 0.70808897 0.29191103
#> [138,] 0.79201827 0.20798173
#> [139,] 0.80762732 0.19237268
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  3.8743284  0.0000000  0.0000000 14.3106245  3.2843465  0.0000000  3.5666374 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.4929843  0.7260934  0.0000000  0.0000000  0.0000000  1.7901695  3.3107351 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  0.0000000  0.0000000  0.7751175  2.8017276  0.0000000  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  4.3291580  2.1391075 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  3.5253582  3.4596466  1.1933332  0.0000000  0.0000000  3.0416406 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  3.6080042  1.9291359  0.0000000  5.3481009  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.3548211  7.4150342  1.6628450  0.0000000  2.0280660  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.7304728  0.0000000  0.7075625  0.0000000  0.0000000  1.9705529  0.0000000 
#>        V60         V7         V8         V9 
#>  3.1714700  0.0000000  0.0000000 11.4529262 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 73 66 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12         V9        V49        V45        V33         V1        V54 
#> 14.3106245 11.4529262  7.4150342  5.3481009  4.3291580  3.8743284  3.7304728 
#>        V42        V15        V36        V37        V48        V21        V13 
#>  3.6080042  3.5666374  3.5253582  3.4596466  3.3548211  3.3107351  3.2843465 
#>        V60        V40        V26        V34        V51        V59        V43 
#>  3.1714700  3.0416406  2.8017276  2.1391075  2.0280660  1.9705529  1.9291359 
#>        V20         V5        V16        V38        V25        V17        V56 
#>  1.7901695  1.6628450  1.4929843  1.1933332  0.7751175  0.7260934  0.7075625 
#>        V10        V11        V14        V18        V19         V2        V22 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V23        V24        V27        V28        V29         V3        V30 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V31        V32        V35        V39         V4        V41        V44 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V46        V47        V50        V52        V53        V55        V57 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58         V6         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058