Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V52>=0.00935 75 14 M (0.81333333 0.18666667)  
#>      4) V13>=0.16265 67  7 M (0.89552239 0.10447761)  
#>        8) V15< 0.6129 59  2 M (0.96610169 0.03389831) *
#>        9) V15>=0.6129 8  3 R (0.37500000 0.62500000) *
#>      5) V13< 0.16265 8  1 R (0.12500000 0.87500000) *
#>    3) V52< 0.00935 64 16 R (0.25000000 0.75000000)  
#>      6) V51>=0.01245 24 10 M (0.58333333 0.41666667)  
#>       12) V21>=0.606 11  1 M (0.90909091 0.09090909) *
#>       13) V21< 0.606 13  4 R (0.30769231 0.69230769) *
#>      7) V51< 0.01245 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 52 M (0.62589928 0.37410072)  
#>    2) V10>=0.18485 87 14 M (0.83908046 0.16091954)  
#>      4) V36< 0.4682 69  3 M (0.95652174 0.04347826) *
#>      5) V36>=0.4682 18  7 R (0.38888889 0.61111111) *
#>    3) V10< 0.18485 52 14 R (0.26923077 0.73076923)  
#>      6) V4>=0.0604 10  1 M (0.90000000 0.10000000) *
#>      7) V4< 0.0604 42  5 R (0.11904762 0.88095238)  
#>       14) V33< 0.2833 7  3 M (0.57142857 0.42857143) *
#>       15) V33>=0.2833 35  1 R (0.02857143 0.97142857) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V9>=0.124 83 23 M (0.72289157 0.27710843)  
#>      4) V37< 0.3563 50  3 M (0.94000000 0.06000000) *
#>      5) V37>=0.3563 33 13 R (0.39393939 0.60606061)  
#>       10) V45>=0.26515 11  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.26515 22  2 R (0.09090909 0.90909091) *
#>    3) V9< 0.124 56 11 R (0.19642857 0.80357143)  
#>      6) V44>=0.19215 24 11 R (0.45833333 0.54166667)  
#>       12) V31< 0.43 10  0 M (1.00000000 0.00000000) *
#>       13) V31>=0.43 14  1 R (0.07142857 0.92857143) *
#>      7) V44< 0.19215 32  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V49>=0.04425 62 14 M (0.77419355 0.22580645)  
#>      4) V44>=0.1434 45  3 M (0.93333333 0.06666667) *
#>      5) V44< 0.1434 17  6 R (0.35294118 0.64705882) *
#>    3) V49< 0.04425 77 19 R (0.24675325 0.75324675)  
#>      6) V58>=0.00925 10  1 M (0.90000000 0.10000000) *
#>      7) V58< 0.00925 67 10 R (0.14925373 0.85074627)  
#>       14) V43>=0.2653 8  2 M (0.75000000 0.25000000) *
#>       15) V43< 0.2653 59  4 R (0.06779661 0.93220339) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 50 R (0.35971223 0.64028777)  
#>    2) V11>=0.17665 77 34 M (0.55844156 0.44155844)  
#>      4) V27>=0.8422 32  1 M (0.96875000 0.03125000) *
#>      5) V27< 0.8422 45 12 R (0.26666667 0.73333333)  
#>       10) V49>=0.0583 10  1 M (0.90000000 0.10000000) *
#>       11) V49< 0.0583 35  3 R (0.08571429 0.91428571) *
#>    3) V11< 0.17665 62  7 R (0.11290323 0.88709677)  
#>      6) V19>=0.83495 7  1 M (0.85714286 0.14285714) *
#>      7) V19< 0.83495 55  1 R (0.01818182 0.98181818) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V51>=0.01555 54  6 M (0.88888889 0.11111111) *
#>    3) V51< 0.01555 85 28 R (0.32941176 0.67058824)  
#>      6) V10>=0.22885 30 10 M (0.66666667 0.33333333)  
#>       12) V1< 0.0404 20  0 M (1.00000000 0.00000000) *
#>       13) V1>=0.0404 10  0 R (0.00000000 1.00000000) *
#>      7) V10< 0.22885 55  8 R (0.14545455 0.85454545)  
#>       14) V31< 0.3186 8  3 M (0.62500000 0.37500000) *
#>       15) V31>=0.3186 47  3 R (0.06382979 0.93617021) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 59 R (0.42446043 0.57553957)  
#>   2) V48>=0.07585 64 16 M (0.75000000 0.25000000)  
#>     4) V36< 0.52295 50  4 M (0.92000000 0.08000000) *
#>     5) V36>=0.52295 14  2 R (0.14285714 0.85714286) *
#>   3) V48< 0.07585 75 11 R (0.14666667 0.85333333)  
#>     6) V59>=0.0072 12  4 M (0.66666667 0.33333333) *
#>     7) V59< 0.0072 63  3 R (0.04761905 0.95238095) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V5>=0.03935 98 32 M (0.67346939 0.32653061)  
#>      4) V23>=0.785 39  2 M (0.94871795 0.05128205) *
#>      5) V23< 0.785 59 29 R (0.49152542 0.50847458)  
#>       10) V25< 0.51305 27  5 M (0.81481481 0.18518519)  
#>         20) V10>=0.197 19  0 M (1.00000000 0.00000000) *
#>         21) V10< 0.197 8  3 R (0.37500000 0.62500000) *
#>       11) V25>=0.51305 32  7 R (0.21875000 0.78125000)  
#>         22) V55< 0.00685 7  1 M (0.85714286 0.14285714) *
#>         23) V55>=0.00685 25  1 R (0.04000000 0.96000000) *
#>    3) V5< 0.03935 41  9 R (0.21951220 0.78048780)  
#>      6) V51>=0.01515 13  4 M (0.69230769 0.30769231) *
#>      7) V51< 0.01515 28  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 50 R (0.35971223 0.64028777)  
#>    2) V45>=0.23645 18  1 M (0.94444444 0.05555556) *
#>    3) V45< 0.23645 121 33 R (0.27272727 0.72727273)  
#>      6) V35< 0.099 12  1 M (0.91666667 0.08333333) *
#>      7) V35>=0.099 109 22 R (0.20183486 0.79816514)  
#>       14) V59>=0.01205 14  4 M (0.71428571 0.28571429) *
#>       15) V59< 0.01205 95 12 R (0.12631579 0.87368421)  
#>         30) V28>=0.92715 14  6 M (0.57142857 0.42857143) *
#>         31) V28< 0.92715 81  4 R (0.04938272 0.95061728) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V52>=0.0233 24  2 M (0.91666667 0.08333333) *
#>    3) V52< 0.0233 115 39 R (0.33913043 0.66086957)  
#>      6) V31< 0.467 41 16 M (0.60975610 0.39024390)  
#>       12) V44>=0.1193 31  6 M (0.80645161 0.19354839)  
#>         24) V30>=0.25285 24  1 M (0.95833333 0.04166667) *
#>         25) V30< 0.25285 7  2 R (0.28571429 0.71428571) *
#>       13) V44< 0.1193 10  0 R (0.00000000 1.00000000) *
#>      7) V31>=0.467 74 14 R (0.18918919 0.81081081)  
#>       14) V16< 0.1644 15  5 M (0.66666667 0.33333333) *
#>       15) V16>=0.1644 59  4 R (0.06779661 0.93220339) *
#> 
#> 
#> $weights
#>  [1] 0.6240720 0.6204556 0.9034564 0.5111766 0.6274507 0.6314966 0.6088633
#>  [8] 0.6541699 0.6687504 0.7615020
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.5729256 4.0384680
#>   [2,] 1.2746255 5.3367681
#>   [3,] 2.3719922 4.2394014
#>   [4,] 3.1837358 3.4276578
#>   [5,] 1.2519522 5.3594414
#>   [6,] 2.4057065 4.2056872
#>   [7,] 2.1513628 4.4600309
#>   [8,] 1.8644320 4.7469617
#>   [9,] 1.4302524 5.1811412
#>  [10,] 2.5301941 4.0811995
#>  [11,] 2.5301941 4.0811995
#>  [12,] 1.9091103 4.7022833
#>  [13,] 2.4092460 4.2021477
#>  [14,] 2.6375263 3.9738674
#>  [15,] 1.1653465 5.4460471
#>  [16,] 1.2782419 5.3331517
#>  [17,] 1.1352487 5.4761450
#>  [18,] 2.4354956 4.1758980
#>  [19,] 1.1799270 5.4314666
#>  [20,] 1.9414290 4.6699646
#>  [21,] 0.6314966 5.9798970
#>  [22,] 1.8678107 4.7435830
#>  [23,] 2.9169107 3.6944830
#>  [24,] 1.7626994 4.8486942
#>  [25,] 1.9944374 4.6169563
#>  [26,] 2.4209152 4.1904785
#>  [27,] 1.8567697 4.7546239
#>  [28,] 1.1653465 5.4460471
#>  [29,] 0.0000000 6.6113936
#>  [30,] 1.7968431 4.8145505
#>  [31,] 1.8114236 4.7999700
#>  [32,] 2.0461297 4.5652640
#>  [33,] 1.1426733 5.4687204
#>  [34,] 1.9414290 4.6699646
#>  [35,] 1.2726787 5.3387150
#>  [36,] 1.9268485 4.6845451
#>  [37,] 1.2293190 5.3820747
#>  [38,] 1.1653465 5.4460471
#>  [39,] 2.0833834 4.5280102
#>  [40,] 1.4156719 5.1957217
#>  [41,] 1.3703653 5.2410283
#>  [42,] 0.6541699 5.9572237
#>  [43,] 2.0844223 4.5269714
#>  [44,] 0.7615020 5.8498916
#>  [45,] 1.1799270 5.4314666
#>  [46,] 1.1200400 5.4913537
#>  [47,] 1.4302524 5.1811412
#>  [48,] 0.6204556 5.9909380
#>  [49,] 1.2746255 5.3367681
#>  [50,] 1.2519522 5.3594414
#>  [51,] 1.9061221 4.7052715
#>  [52,] 2.8091490 3.8022446
#>  [53,] 2.0397439 4.5716497
#>  [54,] 1.9097385 4.7016551
#>  [55,] 1.2293190 5.3820747
#>  [56,] 2.7602261 3.8511675
#>  [57,] 1.1200400 5.4913537
#>  [58,] 2.1327754 4.4786183
#>  [59,] 2.5205850 4.0908086
#>  [60,] 1.8073777 4.8040159
#>  [61,] 1.9131172 4.6982764
#>  [62,] 2.1664896 4.4449040
#>  [63,] 0.9034564 5.7079372
#>  [64,] 2.0234964 4.5878972
#>  [65,] 1.5275284 5.0838652
#>  [66,] 1.8980693 4.7133243
#>  [67,] 1.3703653 5.2410283
#>  [68,] 6.6113936 0.0000000
#>  [69,] 5.3257271 1.2856665
#>  [70,] 4.8040159 1.8073777
#>  [71,] 5.3598709 1.2515227
#>  [72,] 6.1002170 0.5111766
#>  [73,] 5.0804865 1.5309071
#>  [74,] 5.9839429 0.6274507
#>  [75,] 4.6836959 1.9276977
#>  [76,] 4.6981156 1.9132780
#>  [77,] 4.6911205 1.9202731
#>  [78,] 5.9839429 0.6274507
#>  [79,] 5.1967606 1.4146331
#>  [80,] 5.9839429 0.6274507
#>  [81,] 6.6113936 0.0000000
#>  [82,] 5.9426433 0.6687504
#>  [83,] 5.8498916 0.7615020
#>  [84,] 5.3111466 1.3002470
#>  [85,] 6.6113936 0.0000000
#>  [86,] 6.6113936 0.0000000
#>  [87,] 5.3337799 1.2776137
#>  [88,] 5.3185712 1.2928224
#>  [89,] 6.6113936 0.0000000
#>  [90,] 5.3598709 1.2515227
#>  [91,] 4.8486942 1.7626994
#>  [92,] 4.7580027 1.8533910
#>  [93,] 5.3820747 1.2293190
#>  [94,] 5.3820747 1.2293190
#>  [95,] 4.5693098 2.0420838
#>  [96,] 5.9426433 0.6687504
#>  [97,] 5.9839429 0.6274507
#>  [98,] 4.7093174 1.9020762
#>  [99,] 5.3297730 1.2816206
#> [100,] 5.0537673 1.5576263
#> [101,] 5.8498916 0.7615020
#> [102,] 4.8412696 1.7701240
#> [103,] 5.3387150 1.2726787
#> [104,] 3.1390575 3.4723362
#> [105,] 3.2714027 3.3399909
#> [106,] 5.3111466 1.3002470
#> [107,] 4.1985313 2.4128623
#> [108,] 4.0892523 2.5221413
#> [109,] 4.7279048 1.8834888
#> [110,] 5.1811412 1.4302524
#> [111,] 3.9296185 2.6817751
#> [112,] 5.9873216 0.6240720
#> [113,] 5.0804865 1.5309071
#> [114,] 4.7510076 1.8603861
#> [115,] 5.3820747 1.2293190
#> [116,] 5.3784583 1.2329353
#> [117,] 5.3820747 1.2293190
#> [118,] 5.0874816 1.5239120
#> [119,] 5.0391868 1.5722068
#> [120,] 4.0611013 2.5502923
#> [121,] 6.6113936 0.0000000
#> [122,] 5.3710337 1.2403599
#> [123,] 4.2776848 2.3337088
#> [124,] 4.6699646 1.9414290
#> [125,] 5.9572237 0.6541699
#> [126,] 5.1957217 1.4156719
#> [127,] 5.9572237 0.6541699
#> [128,] 5.1957217 1.4156719
#> [129,] 6.1002170 0.5111766
#> [130,] 4.1940949 2.4172988
#> [131,] 5.4761450 1.1352487
#> [132,] 5.8498916 0.7615020
#> [133,] 6.6113936 0.0000000
#> [134,] 5.4913537 1.1200400
#> [135,] 6.6113936 0.0000000
#> [136,] 6.6113936 0.0000000
#> [137,] 4.7146430 1.8967507
#> [138,] 4.4523686 2.1590250
#> [139,] 5.4687204 1.1426733
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.38916540 0.61083460
#>   [2,] 0.19279226 0.80720774
#>   [3,] 0.35877341 0.64122659
#>   [4,] 0.48155291 0.51844709
#>   [5,] 0.18936283 0.81063717
#>   [6,] 0.36387282 0.63612718
#>   [7,] 0.32540231 0.67459769
#>   [8,] 0.28200287 0.71799713
#>   [9,] 0.21633145 0.78366855
#>  [10,] 0.38270209 0.61729791
#>  [11,] 0.38270209 0.61729791
#>  [12,] 0.28876065 0.71123935
#>  [13,] 0.36440819 0.63559181
#>  [14,] 0.39893651 0.60106349
#>  [15,] 0.17626337 0.82373663
#>  [16,] 0.19333925 0.80666075
#>  [17,] 0.17171095 0.82828905
#>  [18,] 0.36837855 0.63162145
#>  [19,] 0.17846873 0.82153127
#>  [20,] 0.29364898 0.70635102
#>  [21,] 0.09551641 0.90448359
#>  [22,] 0.28251391 0.71748609
#>  [23,] 0.44119452 0.55880548
#>  [24,] 0.26661540 0.73338460
#>  [25,] 0.30166671 0.69833329
#>  [26,] 0.36617320 0.63382680
#>  [27,] 0.28084392 0.71915608
#>  [28,] 0.17626337 0.82373663
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.27177979 0.72822021
#>  [31,] 0.27398514 0.72601486
#>  [32,] 0.30948538 0.69051462
#>  [33,] 0.17283395 0.82716605
#>  [34,] 0.29364898 0.70635102
#>  [35,] 0.19249779 0.80750221
#>  [36,] 0.29144363 0.70855637
#>  [37,] 0.18593946 0.81406054
#>  [38,] 0.17626337 0.82373663
#>  [39,] 0.31512016 0.68487984
#>  [40,] 0.21412610 0.78587390
#>  [41,] 0.20727330 0.79272670
#>  [42,] 0.09894584 0.90105416
#>  [43,] 0.31527729 0.68472271
#>  [44,] 0.11518026 0.88481974
#>  [45,] 0.17846873 0.82153127
#>  [46,] 0.16941057 0.83058943
#>  [47,] 0.21633145 0.78366855
#>  [48,] 0.09384642 0.90615358
#>  [49,] 0.19279226 0.80720774
#>  [50,] 0.18936283 0.81063717
#>  [51,] 0.28830867 0.71169133
#>  [52,] 0.42489514 0.57510486
#>  [53,] 0.30851951 0.69148049
#>  [54,] 0.28885567 0.71114433
#>  [55,] 0.18593946 0.81406054
#>  [56,] 0.41749535 0.58250465
#>  [57,] 0.16941057 0.83058943
#>  [58,] 0.32259089 0.67740911
#>  [59,] 0.38124867 0.61875133
#>  [60,] 0.27337318 0.72662682
#>  [61,] 0.28936671 0.71063329
#>  [62,] 0.32769031 0.67230969
#>  [63,] 0.13665143 0.86334857
#>  [64,] 0.30606200 0.69393800
#>  [65,] 0.23104485 0.76895515
#>  [66,] 0.28709065 0.71290935
#>  [67,] 0.20727330 0.79272670
#>  [68,] 1.00000000 0.00000000
#>  [69,] 0.80553775 0.19446225
#>  [70,] 0.72662682 0.27337318
#>  [71,] 0.81070213 0.18929787
#>  [72,] 0.92268247 0.07731753
#>  [73,] 0.76844411 0.23155589
#>  [74,] 0.90509554 0.09490446
#>  [75,] 0.70842793 0.29157207
#>  [76,] 0.71060897 0.28939103
#>  [77,] 0.70955093 0.29044907
#>  [78,] 0.90509554 0.09490446
#>  [79,] 0.78603103 0.21396897
#>  [80,] 0.90509554 0.09490446
#>  [81,] 1.00000000 0.00000000
#>  [82,] 0.89884881 0.10115119
#>  [83,] 0.88481974 0.11518026
#>  [84,] 0.80333239 0.19666761
#>  [85,] 1.00000000 0.00000000
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.80675577 0.19324423
#>  [88,] 0.80445539 0.19554461
#>  [89,] 1.00000000 0.00000000
#>  [90,] 0.81070213 0.18929787
#>  [91,] 0.73338460 0.26661540
#>  [92,] 0.71966713 0.28033287
#>  [93,] 0.81406054 0.18593946
#>  [94,] 0.81406054 0.18593946
#>  [95,] 0.69112658 0.30887342
#>  [96,] 0.89884881 0.10115119
#>  [97,] 0.90509554 0.09490446
#>  [98,] 0.71230328 0.28769672
#>  [99,] 0.80614970 0.19385030
#> [100,] 0.76440273 0.23559727
#> [101,] 0.88481974 0.11518026
#> [102,] 0.73226160 0.26773840
#> [103,] 0.80750221 0.19249779
#> [104,] 0.47479513 0.52520487
#> [105,] 0.49481288 0.50518712
#> [106,] 0.80333239 0.19666761
#> [107,] 0.63504482 0.36495518
#> [108,] 0.61851593 0.38148407
#> [109,] 0.71511470 0.28488530
#> [110,] 0.78366855 0.21633145
#> [111,] 0.59437068 0.40562932
#> [112,] 0.90560659 0.09439341
#> [113,] 0.76844411 0.23155589
#> [114,] 0.71860909 0.28139091
#> [115,] 0.81406054 0.18593946
#> [116,] 0.81351355 0.18648645
#> [117,] 0.81406054 0.18593946
#> [118,] 0.76950215 0.23049785
#> [119,] 0.76219737 0.23780263
#> [120,] 0.61425798 0.38574202
#> [121,] 1.00000000 0.00000000
#> [122,] 0.81239055 0.18760945
#> [123,] 0.64701712 0.35298288
#> [124,] 0.70635102 0.29364898
#> [125,] 0.90105416 0.09894584
#> [126,] 0.78587390 0.21412610
#> [127,] 0.90105416 0.09894584
#> [128,] 0.78587390 0.21412610
#> [129,] 0.92268247 0.07731753
#> [130,] 0.63437379 0.36562621
#> [131,] 0.82828905 0.17171095
#> [132,] 0.88481974 0.11518026
#> [133,] 1.00000000 0.00000000
#> [134,] 0.83058943 0.16941057
#> [135,] 1.00000000 0.00000000
#> [136,] 1.00000000 0.00000000
#> [137,] 0.71310880 0.28689120
#> [138,] 0.67343874 0.32656126
#> [139,] 0.82716605 0.17283395
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "R" "R" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 2.6375389 7.0958577 2.6800105 0.0000000 1.6589293 0.0000000 0.9623774 2.0462888 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 1.7181116 0.0000000 0.0000000 0.8425651 0.0000000 2.0114000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 2.1323494 0.0000000 3.6236633 1.3629813 0.0000000 0.0000000 1.1697043 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 5.9249165 0.0000000 0.6682173 0.0000000 2.3142659 4.3077561 3.3551649 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 1.9148065 0.0000000 0.0000000 0.0000000 1.0499897 5.3077142 6.3924771 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 0.0000000 4.7949050 5.0871324 2.4413756 0.0000000 7.5014294 7.4442873 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 0.0000000 1.4965586 0.0000000 0.0000000 1.5705818 3.2408248 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 5.2458194 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V51       V52       V10       V45       V31       V44        V9       V49 
#> 7.5014294 7.4442873 7.0958577 6.3924771 5.9249165 5.3077142 5.2458194 5.0871324 
#>       V48       V36       V27       V37       V59       V11        V1        V5 
#> 4.7949050 4.3077561 3.6236633 3.3551649 3.2408248 2.6800105 2.6375389 2.4413756 
#>       V35       V25       V16       V23        V4       V19       V13       V58 
#> 2.3142659 2.1323494 2.0462888 2.0114000 1.9148065 1.7181116 1.6589293 1.5705818 
#>       V55       V28       V30       V43       V15       V21       V33       V12 
#> 1.4965586 1.3629813 1.1697043 1.0499897 0.9623774 0.8425651 0.6682173 0.0000000 
#>       V14       V17       V18        V2       V20       V22       V24       V26 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V29        V3       V32       V34       V38       V39       V40       V41 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V42       V46       V47       V50       V53       V54       V56       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>        V6       V60        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768