Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 61 R (0.43884892 0.56115108)  
#>   2) V11>=0.1998 75 22 M (0.70666667 0.29333333)  
#>     4) V52>=0.0069 64 12 M (0.81250000 0.18750000)  
#>       8) V29>=0.447 51  3 M (0.94117647 0.05882353) *
#>       9) V29< 0.447 13  4 R (0.30769231 0.69230769) *
#>     5) V52< 0.0069 11  1 R (0.09090909 0.90909091) *
#>   3) V11< 0.1998 64  8 R (0.12500000 0.87500000)  
#>     6) V1>=0.0392 10  3 M (0.70000000 0.30000000) *
#>     7) V1< 0.0392 54  1 R (0.01851852 0.98148148) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V49>=0.03685 76 22 M (0.71052632 0.28947368)  
#>      4) V51>=0.01285 54  6 M (0.88888889 0.11111111) *
#>      5) V51< 0.01285 22  6 R (0.27272727 0.72727273)  
#>       10) V28>=0.75935 7  2 M (0.71428571 0.28571429) *
#>       11) V28< 0.75935 15  1 R (0.06666667 0.93333333) *
#>    3) V49< 0.03685 63 13 R (0.20634921 0.79365079)  
#>      6) V28< 0.43995 9  1 M (0.88888889 0.11111111) *
#>      7) V28>=0.43995 54  5 R (0.09259259 0.90740741) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V43>=0.35065 29  3 M (0.89655172 0.10344828) *
#>    3) V43< 0.35065 110 41 R (0.37272727 0.62727273)  
#>      6) V36< 0.48825 66 27 M (0.59090909 0.40909091)  
#>       12) V1>=0.0157 49 12 M (0.75510204 0.24489796)  
#>         24) V23>=0.7542 28  2 M (0.92857143 0.07142857) *
#>         25) V23< 0.7542 21 10 M (0.52380952 0.47619048)  
#>           50) V28< 0.53015 9  1 M (0.88888889 0.11111111) *
#>           51) V28>=0.53015 12  3 R (0.25000000 0.75000000) *
#>       13) V1< 0.0157 17  2 R (0.11764706 0.88235294) *
#>      7) V36>=0.48825 44  2 R (0.04545455 0.95454545) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V11>=0.1534 95 25 M (0.73684211 0.26315789)  
#>      4) V40< 0.44795 75 10 M (0.86666667 0.13333333)  
#>        8) V7< 0.20865 68  4 M (0.94117647 0.05882353) *
#>        9) V7>=0.20865 7  1 R (0.14285714 0.85714286) *
#>      5) V40>=0.44795 20  5 R (0.25000000 0.75000000)  
#>       10) V21>=0.80365 8  3 M (0.62500000 0.37500000) *
#>       11) V21< 0.80365 12  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1534 44  7 R (0.15909091 0.84090909)  
#>      6) V53< 0.00535 11  5 M (0.54545455 0.45454545) *
#>      7) V53>=0.00535 33  1 R (0.03030303 0.96969697) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V9>=0.15095 57 16 M (0.71929825 0.28070175)  
#>      4) V17< 0.4374 29  1 M (0.96551724 0.03448276) *
#>      5) V17>=0.4374 28 13 R (0.46428571 0.53571429)  
#>       10) V36< 0.2923 17  4 M (0.76470588 0.23529412) *
#>       11) V36>=0.2923 11  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.15095 82 21 R (0.25609756 0.74390244)  
#>      6) V4>=0.0611 13  1 M (0.92307692 0.07692308) *
#>      7) V4< 0.0611 69  9 R (0.13043478 0.86956522)  
#>       14) V28>=0.96735 10  1 M (0.90000000 0.10000000) *
#>       15) V28< 0.96735 59  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V51>=0.01365 78 20 M (0.74358974 0.25641026)  
#>      4) V12>=0.14575 62  8 M (0.87096774 0.12903226)  
#>        8) V7< 0.21185 53  2 M (0.96226415 0.03773585) *
#>        9) V7>=0.21185 9  3 R (0.33333333 0.66666667) *
#>      5) V12< 0.14575 16  4 R (0.25000000 0.75000000) *
#>    3) V51< 0.01365 61 17 R (0.27868852 0.72131148)  
#>      6) V38>=0.42275 23  9 M (0.60869565 0.39130435)  
#>       12) V31< 0.53425 13  0 M (1.00000000 0.00000000) *
#>       13) V31>=0.53425 10  1 R (0.10000000 0.90000000) *
#>      7) V38< 0.42275 38  3 R (0.07894737 0.92105263) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.4316547 0.5683453)  
#>    2) V45>=0.26365 37  6 M (0.8378378 0.1621622)  
#>      4) V40< 0.58555 30  0 M (1.0000000 0.0000000) *
#>      5) V40>=0.58555 7  1 R (0.1428571 0.8571429) *
#>    3) V45< 0.26365 102 29 R (0.2843137 0.7156863)  
#>      6) V12>=0.1922 65 29 R (0.4461538 0.5538462)  
#>       12) V31< 0.4457 21  4 M (0.8095238 0.1904762)  
#>         24) V31>=0.29095 14  0 M (1.0000000 0.0000000) *
#>         25) V31< 0.29095 7  3 R (0.4285714 0.5714286) *
#>       13) V31>=0.4457 44 12 R (0.2727273 0.7272727)  
#>         26) V57< 0.00585 10  2 M (0.8000000 0.2000000) *
#>         27) V57>=0.00585 34  4 R (0.1176471 0.8823529)  
#>           54) V18< 0.277 7  3 M (0.5714286 0.4285714) *
#>           55) V18>=0.277 27  0 R (0.0000000 1.0000000) *
#>      7) V12< 0.1922 37  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V45>=0.38545 29  0 M (1.00000000 0.00000000) *
#>    3) V45< 0.38545 110 36 R (0.32727273 0.67272727)  
#>      6) V42>=0.26545 48 20 M (0.58333333 0.41666667)  
#>       12) V36< 0.5274 32  5 M (0.84375000 0.15625000) *
#>       13) V36>=0.5274 16  1 R (0.06250000 0.93750000) *
#>      7) V42< 0.26545 62  8 R (0.12903226 0.87096774)  
#>       14) V37< 0.12475 9  4 M (0.55555556 0.44444444) *
#>       15) V37>=0.12475 53  3 R (0.05660377 0.94339623) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V49>=0.05255 62 16 M (0.74193548 0.25806452)  
#>      4) V30>=0.63865 31  0 M (1.00000000 0.00000000) *
#>      5) V30< 0.63865 31 15 R (0.48387097 0.51612903)  
#>       10) V36< 0.2913 11  0 M (1.00000000 0.00000000) *
#>       11) V36>=0.2913 20  4 R (0.20000000 0.80000000) *
#>    3) V49< 0.05255 77 23 R (0.29870130 0.70129870)  
#>      6) V40< 0.10525 12  1 M (0.91666667 0.08333333) *
#>      7) V40>=0.10525 65 12 R (0.18461538 0.81538462)  
#>       14) V20>=0.88045 11  4 M (0.63636364 0.36363636) *
#>       15) V20< 0.88045 54  5 R (0.09259259 0.90740741)  
#>         30) V57< 0.0029 7  2 M (0.71428571 0.28571429) *
#>         31) V57>=0.0029 47  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.5395683 0.4604317)  
#>    2) V56>=0.01015 50  6 M (0.8800000 0.1200000)  
#>      4) V38>=0.24595 41  0 M (1.0000000 0.0000000) *
#>      5) V38< 0.24595 9  3 R (0.3333333 0.6666667) *
#>    3) V56< 0.01015 89 31 R (0.3483146 0.6516854)  
#>      6) V51>=0.01545 37 14 M (0.6216216 0.3783784)  
#>       12) V41< 0.26775 19  0 M (1.0000000 0.0000000) *
#>       13) V41>=0.26775 18  4 R (0.2222222 0.7777778) *
#>      7) V51< 0.01545 52  8 R (0.1538462 0.8461538) *
#> 
#> 
#> $weights
#>  [1] 0.7586613 0.8405564 0.7094110 0.5948734 0.9071553 0.7267262 0.7902898
#>  [8] 0.6697552 0.5080125 0.5539871
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.4997007 5.5597275
#>   [2,] 1.3126484 5.7467798
#>   [3,] 2.8319112 4.2275170
#>   [4,] 1.3945435 5.6648847
#>   [5,] 1.1028859 5.9565424
#>   [6,] 0.5080125 6.5514157
#>   [7,] 1.0619996 5.9974286
#>   [8,] 1.3945435 5.6648847
#>   [9,] 1.5170160 5.5424122
#>  [10,] 0.6697552 6.3896730
#>  [11,] 1.7477117 5.3117165
#>  [12,] 2.2756773 4.7837509
#>  [13,] 2.4744380 4.5849902
#>  [14,] 0.5948734 6.4645549
#>  [15,] 2.9254794 4.1339488
#>  [16,] 2.1438244 4.9156038
#>  [17,] 2.7737097 4.2857185
#>  [18,] 1.9434422 5.1159860
#>  [19,] 2.4993673 4.5600609
#>  [20,] 1.3851631 5.6742651
#>  [21,] 2.0100412 5.0493870
#>  [22,] 1.8931756 5.1662526
#>  [23,] 0.5948734 6.4645549
#>  [24,] 0.5080125 6.5514157
#>  [25,] 2.8799311 4.1794971
#>  [26,] 1.4354297 5.6239985
#>  [27,] 0.0000000 7.0594282
#>  [28,] 0.5948734 6.4645549
#>  [29,] 0.0000000 7.0594282
#>  [30,] 1.3042843 5.7551439
#>  [31,] 2.5640283 4.4953999
#>  [32,] 0.0000000 7.0594282
#>  [33,] 0.5948734 6.4645549
#>  [34,] 1.4853876 5.5740407
#>  [35,] 1.3042843 5.7551439
#>  [36,] 1.1028859 5.9565424
#>  [37,] 0.7094110 6.3500173
#>  [38,] 1.3791662 5.6802621
#>  [39,] 0.5948734 6.4645549
#>  [40,] 2.2257195 4.8337087
#>  [41,] 1.1488605 5.9105678
#>  [42,] 0.9071553 6.1522729
#>  [43,] 0.5080125 6.5514157
#>  [44,] 1.5020287 5.5573995
#>  [45,] 0.9071553 6.1522729
#>  [46,] 0.5948734 6.4645549
#>  [47,] 1.8755867 5.1838415
#>  [48,] 0.0000000 7.0594282
#>  [49,] 2.0077132 5.0517150
#>  [50,] 0.7902898 6.2691384
#>  [51,] 1.7726411 5.2867872
#>  [52,] 1.1777677 5.8816605
#>  [53,] 0.5080125 6.5514157
#>  [54,] 1.1028859 5.9565424
#>  [55,] 1.2633981 5.7960301
#>  [56,] 1.7714106 5.2880176
#>  [57,] 1.3791662 5.6802621
#>  [58,] 2.2197226 4.8397057
#>  [59,] 1.7317548 5.3276734
#>  [60,] 1.2983023 5.7611259
#>  [61,] 2.7687025 4.2907257
#>  [62,] 2.0100412 5.0493870
#>  [63,] 2.0945741 4.9648541
#>  [64,] 2.8811949 4.1782334
#>  [65,] 0.8405564 6.2188718
#>  [66,] 1.4853876 5.5740407
#>  [67,] 1.4354297 5.6239985
#>  [68,] 1.5992177 5.4602105
#>  [69,] 1.5769105 5.4825177
#>  [70,] 1.4600450 5.5993832
#>  [71,] 0.5080125 6.5514157
#>  [72,] 4.5600609 2.4993673
#>  [73,] 4.3808623 2.6785659
#>  [74,] 4.9381585 2.1212698
#>  [75,] 4.5690121 2.4904161
#>  [76,] 4.8407219 2.2187063
#>  [77,] 4.1698692 2.8895590
#>  [78,] 4.6033218 2.4561064
#>  [79,] 4.9216007 2.1378275
#>  [80,] 4.4075913 2.6518369
#>  [81,] 5.1159860 1.9434422
#>  [82,] 5.5982858 1.4611425
#>  [83,] 4.9348494 2.1245788
#>  [84,] 5.0024646 2.0569636
#>  [85,] 4.8899723 2.1694559
#>  [86,] 6.5054411 0.5539871
#>  [87,] 5.7927544 1.2666738
#>  [88,] 6.3896730 0.6697552
#>  [89,] 5.7108593 1.3485689
#>  [90,] 6.3896730 0.6697552
#>  [91,] 6.3896730 0.6697552
#>  [92,] 5.5993832 1.4600450
#>  [93,] 5.0024646 2.0569636
#>  [94,] 4.8330012 2.2264270
#>  [95,] 5.0833434 1.9760848
#>  [96,] 6.2691384 0.7902898
#>  [97,] 4.3586625 2.7007658
#>  [98,] 6.2188718 0.8405564
#>  [99,] 5.2387673 1.8206609
#> [100,] 5.2298161 1.8296121
#> [101,] 7.0594282 0.0000000
#> [102,] 7.0594282 0.0000000
#> [103,] 6.3500173 0.7094110
#> [104,] 5.6648847 1.3945435
#> [105,] 4.4251802 2.6342480
#> [106,] 6.5054411 0.5539871
#> [107,] 6.5054411 0.5539871
#> [108,] 4.2867348 2.7726934
#> [109,] 4.9285306 2.1308977
#> [110,] 4.7837509 2.2756773
#> [111,] 5.7787149 1.2807134
#> [112,] 5.1838415 1.8755867
#> [113,] 5.9105678 1.1488605
#> [114,] 5.7787149 1.2807134
#> [115,] 5.1202780 1.9391503
#> [116,] 5.5982858 1.4611425
#> [117,] 5.2387673 1.8206609
#> [118,] 3.9258459 3.1335823
#> [119,] 4.3566172 2.7028110
#> [120,] 4.9951295 2.0642987
#> [121,] 4.2497169 2.8097114
#> [122,] 5.0693039 1.9901243
#> [123,] 7.0594282 0.0000000
#> [124,] 7.0594282 0.0000000
#> [125,] 7.0594282 0.0000000
#> [126,] 6.5054411 0.5539871
#> [127,] 5.8420048 1.2174235
#> [128,] 5.8420048 1.2174235
#> [129,] 5.8420048 1.2174235
#> [130,] 5.5094609 1.5499674
#> [131,] 5.0693039 1.9901243
#> [132,] 6.3500173 0.7094110
#> [133,] 5.7960301 1.2633981
#> [134,] 5.2707024 1.7887259
#> [135,] 5.8356859 1.2237423
#> [136,] 5.0200535 2.0393747
#> [137,] 7.0594282 0.0000000
#> [138,] 4.3995487 2.6598795
#> [139,] 5.5491166 1.5103116
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.21243941 0.78756059
#>   [2,] 0.18594260 0.81405740
#>   [3,] 0.40115305 0.59884695
#>   [4,] 0.19754341 0.80245659
#>   [5,] 0.15622878 0.84377122
#>   [6,] 0.07196227 0.92803773
#>   [7,] 0.15043706 0.84956294
#>   [8,] 0.19754341 0.80245659
#>   [9,] 0.21489220 0.78510780
#>  [10,] 0.09487386 0.90512614
#>  [11,] 0.24757129 0.75242871
#>  [12,] 0.32236001 0.67763999
#>  [13,] 0.35051535 0.64948465
#>  [14,] 0.08426651 0.91573349
#>  [15,] 0.41440742 0.58559258
#>  [16,] 0.30368245 0.69631755
#>  [17,] 0.39290855 0.60709145
#>  [18,] 0.27529740 0.72470260
#>  [19,] 0.35404670 0.64595330
#>  [20,] 0.19621463 0.80378537
#>  [21,] 0.28473144 0.71526856
#>  [22,] 0.26817691 0.73182309
#>  [23,] 0.08426651 0.91573349
#>  [24,] 0.07196227 0.92803773
#>  [25,] 0.40795529 0.59204471
#>  [26,] 0.20333513 0.79666487
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.08426651 0.91573349
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.18475778 0.81524222
#>  [31,] 0.36320623 0.63679377
#>  [32,] 0.00000000 1.00000000
#>  [33,] 0.08426651 0.91573349
#>  [34,] 0.21041188 0.78958812
#>  [35,] 0.18475778 0.81524222
#>  [36,] 0.15622878 0.84377122
#>  [37,] 0.10049128 0.89950872
#>  [38,] 0.19536514 0.80463486
#>  [39,] 0.08426651 0.91573349
#>  [40,] 0.31528326 0.68471674
#>  [41,] 0.16274129 0.83725871
#>  [42,] 0.12850267 0.87149733
#>  [43,] 0.07196227 0.92803773
#>  [44,] 0.21276917 0.78723083
#>  [45,] 0.12850267 0.87149733
#>  [46,] 0.08426651 0.91573349
#>  [47,] 0.26568536 0.73431464
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.28440168 0.71559832
#>  [50,] 0.11194813 0.88805187
#>  [51,] 0.25110264 0.74889736
#>  [52,] 0.16683613 0.83316387
#>  [53,] 0.07196227 0.92803773
#>  [54,] 0.15622878 0.84377122
#>  [55,] 0.17896606 0.82103394
#>  [56,] 0.25092834 0.74907166
#>  [57,] 0.19536514 0.80463486
#>  [58,] 0.31443376 0.68556624
#>  [59,] 0.24531092 0.75468908
#>  [60,] 0.18391040 0.81608960
#>  [61,] 0.39219926 0.60780074
#>  [62,] 0.28473144 0.71526856
#>  [63,] 0.29670591 0.70329409
#>  [64,] 0.40813431 0.59186569
#>  [65,] 0.11906862 0.88093138
#>  [66,] 0.21041188 0.78958812
#>  [67,] 0.20333513 0.79666487
#>  [68,] 0.22653644 0.77346356
#>  [69,] 0.22337653 0.77662347
#>  [70,] 0.20682199 0.79317801
#>  [71,] 0.07196227 0.92803773
#>  [72,] 0.64595330 0.35404670
#>  [73,] 0.62056900 0.37943100
#>  [74,] 0.69951252 0.30048748
#>  [75,] 0.64722127 0.35277873
#>  [76,] 0.68571020 0.31428980
#>  [77,] 0.59068088 0.40931912
#>  [78,] 0.65208139 0.34791861
#>  [79,] 0.69716705 0.30283295
#>  [80,] 0.62435528 0.37564472
#>  [81,] 0.72470260 0.27529740
#>  [82,] 0.79302255 0.20697745
#>  [83,] 0.69904378 0.30095622
#>  [84,] 0.70862179 0.29137821
#>  [85,] 0.69268673 0.30731327
#>  [86,] 0.92152521 0.07847479
#>  [87,] 0.82056991 0.17943009
#>  [88,] 0.90512614 0.09487386
#>  [89,] 0.80896910 0.19103090
#>  [90,] 0.90512614 0.09487386
#>  [91,] 0.90512614 0.09487386
#>  [92,] 0.79317801 0.20682199
#>  [93,] 0.70862179 0.29137821
#>  [94,] 0.68461653 0.31538347
#>  [95,] 0.72007864 0.27992136
#>  [96,] 0.88805187 0.11194813
#>  [97,] 0.61742429 0.38257571
#>  [98,] 0.88093138 0.11906862
#>  [99,] 0.74209513 0.25790487
#> [100,] 0.74082715 0.25917285
#> [101,] 1.00000000 0.00000000
#> [102,] 1.00000000 0.00000000
#> [103,] 0.89950872 0.10049128
#> [104,] 0.80245659 0.19754341
#> [105,] 0.62684683 0.37315317
#> [106,] 0.92152521 0.07847479
#> [107,] 0.92152521 0.07847479
#> [108,] 0.60723541 0.39276459
#> [109,] 0.69814869 0.30185131
#> [110,] 0.67763999 0.32236001
#> [111,] 0.81858115 0.18141885
#> [112,] 0.73431464 0.26568536
#> [113,] 0.83725871 0.16274129
#> [114,] 0.81858115 0.18141885
#> [115,] 0.72531058 0.27468942
#> [116,] 0.79302255 0.20697745
#> [117,] 0.74209513 0.25790487
#> [118,] 0.55611386 0.44388614
#> [119,] 0.61713457 0.38286543
#> [120,] 0.70758273 0.29241727
#> [121,] 0.60199165 0.39800835
#> [122,] 0.71808987 0.28191013
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 0.92152521 0.07847479
#> [127,] 0.82754645 0.17245355
#> [128,] 0.82754645 0.17245355
#> [129,] 0.82754645 0.17245355
#> [130,] 0.78044010 0.21955990
#> [131,] 0.71808987 0.28191013
#> [132,] 0.89950872 0.10049128
#> [133,] 0.82103394 0.17896606
#> [134,] 0.74661887 0.25338113
#> [135,] 0.82665135 0.17334865
#> [136,] 0.71111333 0.28888667
#> [137,] 1.00000000 0.00000000
#> [138,] 0.62321601 0.37678399
#> [139,] 0.78605752 0.21394248
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 3.7616324 0.0000000 8.4409767 4.1382421 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 1.8473137 0.8161955 0.0000000 0.0000000 0.7810533 0.6346391 0.0000000 0.7935594 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 0.0000000 7.7190910 1.7945438 0.0000000 1.1935036 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 4.4201936 0.0000000 0.0000000 0.0000000 0.0000000 8.9801282 0.7298960 2.6965090 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 3.5474445 5.4765157 1.7625285 2.1278691 2.5418292 0.0000000 7.6986116 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 0.0000000 0.0000000 6.1380165 0.0000000 0.0000000 7.3885210 2.1098058 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.7410533 0.0000000 0.0000000 2.8526953 2.5163330 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 2.6274248 0.0000000 3.7238742 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 68 71 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V36       V11       V28       V45       V51       V49       V40       V31 
#> 8.9801282 8.4409767 7.7190910 7.6986116 7.3885210 6.1380165 5.4765157 4.4201936 
#>       V12        V1        V9        V4       V56       V38        V7       V43 
#> 4.1382421 3.7616324 3.7238742 3.5474445 2.8526953 2.6965090 2.6274248 2.5418292 
#>       V57       V42       V52       V17       V29       V41       V30       V18 
#> 2.5163330 2.1278691 2.1098058 1.8473137 1.7945438 1.7625285 1.1935036 0.8161955 
#>       V23       V20       V53       V37       V21       V10       V13       V14 
#> 0.7935594 0.7810533 0.7410533 0.7298960 0.6346391 0.0000000 0.0000000 0.0000000 
#>       V15       V16       V19        V2       V22       V24       V25       V26 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V27        V3       V32       V33       V34       V35       V39       V44 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V46       V47       V48        V5       V50       V54       V55       V58 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V59        V6       V60        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623