Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V12>=0.22505 72 12 M (0.83333333 0.16666667)  
#>      4) V9>=0.10645 65  6 M (0.90769231 0.09230769) *
#>      5) V9< 0.10645 7  1 R (0.14285714 0.85714286) *
#>    3) V12< 0.22505 67 16 R (0.23880597 0.76119403)  
#>      6) V26< 0.514 12  3 M (0.75000000 0.25000000) *
#>      7) V26>=0.514 55  7 R (0.12727273 0.87272727)  
#>       14) V53>=0.01585 9  3 M (0.66666667 0.33333333) *
#>       15) V53< 0.01585 46  1 R (0.02173913 0.97826087) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 60 R (0.43165468 0.56834532)  
#>   2) V10>=0.16625 67 18 M (0.73134328 0.26865672)  
#>     4) V47>=0.06475 54  7 M (0.87037037 0.12962963) *
#>     5) V47< 0.06475 13  2 R (0.15384615 0.84615385) *
#>   3) V10< 0.16625 72 11 R (0.15277778 0.84722222)  
#>     6) V45>=0.1728 14  5 M (0.64285714 0.35714286) *
#>     7) V45< 0.1728 58  2 R (0.03448276 0.96551724) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V44>=0.24985 42 11 M (0.73809524 0.26190476)  
#>      4) V9>=0.10435 25  1 M (0.96000000 0.04000000) *
#>      5) V9< 0.10435 17  7 R (0.41176471 0.58823529) *
#>    3) V44< 0.24985 97 27 R (0.27835052 0.72164948)  
#>      6) V35< 0.1909 25  8 M (0.68000000 0.32000000)  
#>       12) V8< 0.17625 18  2 M (0.88888889 0.11111111) *
#>       13) V8>=0.17625 7  1 R (0.14285714 0.85714286) *
#>      7) V35>=0.1909 72 10 R (0.13888889 0.86111111)  
#>       14) V45< 0.0669 11  5 M (0.54545455 0.45454545) *
#>       15) V45>=0.0669 61  4 R (0.06557377 0.93442623)  
#>         30) V55< 0.00375 7  3 M (0.57142857 0.42857143) *
#>         31) V55>=0.00375 54  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V54>=0.01595 30  4 M (0.86666667 0.13333333)  
#>      4) V43>=0.2422 23  0 M (1.00000000 0.00000000) *
#>      5) V43< 0.2422 7  3 R (0.42857143 0.57142857) *
#>    3) V54< 0.01595 109 33 R (0.30275229 0.69724771)  
#>      6) V9>=0.1608 48 20 M (0.58333333 0.41666667)  
#>       12) V15< 0.5312 34  7 M (0.79411765 0.20588235)  
#>         24) V14>=0.1866 27  2 M (0.92592593 0.07407407) *
#>         25) V14< 0.1866 7  2 R (0.28571429 0.71428571) *
#>       13) V15>=0.5312 14  1 R (0.07142857 0.92857143) *
#>      7) V9< 0.1608 61  5 R (0.08196721 0.91803279)  
#>       14) V24>=0.90655 7  3 M (0.57142857 0.42857143) *
#>       15) V24< 0.90655 54  1 R (0.01851852 0.98148148) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V31< 0.467 65 15 M (0.76923077 0.23076923)  
#>      4) V22>=0.72395 45  1 M (0.97777778 0.02222222) *
#>      5) V22< 0.72395 20  6 R (0.30000000 0.70000000)  
#>       10) V24< 0.54155 8  2 M (0.75000000 0.25000000) *
#>       11) V24>=0.54155 12  0 R (0.00000000 1.00000000) *
#>    3) V31>=0.467 74 19 R (0.25675676 0.74324324)  
#>      6) V28>=0.9907 9  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.9907 65 10 R (0.15384615 0.84615385)  
#>       14) V49>=0.09695 7  1 M (0.85714286 0.14285714) *
#>       15) V49< 0.09695 58  4 R (0.06896552 0.93103448) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V52>=0.00885 73 24 M (0.67123288 0.32876712)  
#>      4) V11>=0.1221 55  7 M (0.87272727 0.12727273)  
#>        8) V36< 0.52295 47  1 M (0.97872340 0.02127660) *
#>        9) V36>=0.52295 8  2 R (0.25000000 0.75000000) *
#>      5) V11< 0.1221 18  1 R (0.05555556 0.94444444) *
#>    3) V52< 0.00885 66 13 R (0.19696970 0.80303030)  
#>      6) V47>=0.1255 11  3 M (0.72727273 0.27272727) *
#>      7) V47< 0.1255 55  5 R (0.09090909 0.90909091)  
#>       14) V1>=0.0411 9  4 M (0.55555556 0.44444444) *
#>       15) V1< 0.0411 46  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V45>=0.21745 36  3 M (0.91666667 0.08333333) *
#>    3) V45< 0.21745 103 28 R (0.27184466 0.72815534)  
#>      6) V52>=0.00885 51 24 R (0.47058824 0.52941176)  
#>       12) V12>=0.1451 31  8 M (0.74193548 0.25806452)  
#>         24) V36< 0.47405 24  2 M (0.91666667 0.08333333) *
#>         25) V36>=0.47405 7  1 R (0.14285714 0.85714286) *
#>       13) V12< 0.1451 20  1 R (0.05000000 0.95000000) *
#>      7) V52< 0.00885 52  4 R (0.07692308 0.92307692) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V12>=0.1466 103 42 M (0.59223301 0.40776699)  
#>      4) V23>=0.74575 46  4 M (0.91304348 0.08695652)  
#>        8) V21>=0.6496 39  0 M (1.00000000 0.00000000) *
#>        9) V21< 0.6496 7  3 R (0.42857143 0.57142857) *
#>      5) V23< 0.74575 57 19 R (0.33333333 0.66666667)  
#>       10) V43>=0.37395 9  0 M (1.00000000 0.00000000) *
#>       11) V43< 0.37395 48 10 R (0.20833333 0.79166667)  
#>         22) V16< 0.2129 8  2 M (0.75000000 0.25000000) *
#>         23) V16>=0.2129 40  4 R (0.10000000 0.90000000) *
#>    3) V12< 0.1466 36  3 R (0.08333333 0.91666667) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V57< 0.00575 53 10 M (0.81132075 0.18867925)  
#>      4) V14>=0.14275 44  2 M (0.95454545 0.04545455) *
#>      5) V14< 0.14275 9  1 R (0.11111111 0.88888889) *
#>    3) V57>=0.00575 86 31 R (0.36046512 0.63953488)  
#>      6) V59>=0.00715 35 12 M (0.65714286 0.34285714)  
#>       12) V20>=0.68545 20  0 M (1.00000000 0.00000000) *
#>       13) V20< 0.68545 15  3 R (0.20000000 0.80000000) *
#>      7) V59< 0.00715 51  8 R (0.15686275 0.84313725)  
#>       14) V59< 0.0023 7  2 M (0.71428571 0.28571429) *
#>       15) V59>=0.0023 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V27>=0.89995 36  7 M (0.80555556 0.19444444)  
#>      4) V12>=0.21345 29  0 M (1.00000000 0.00000000) *
#>      5) V12< 0.21345 7  0 R (0.00000000 1.00000000) *
#>    3) V27< 0.89995 103 32 R (0.31067961 0.68932039)  
#>      6) V4>=0.0606 33 12 M (0.63636364 0.36363636)  
#>       12) V59>=0.0065 22  1 M (0.95454545 0.04545455) *
#>       13) V59< 0.0065 11  0 R (0.00000000 1.00000000) *
#>      7) V4< 0.0606 70 11 R (0.15714286 0.84285714)  
#>       14) V27< 0.3949 11  4 M (0.63636364 0.36363636) *
#>       15) V27>=0.3949 59  4 R (0.06779661 0.93220339) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.6971743 0.4819460 0.6153816 0.7399534 0.6810222 0.6477782
#>  [8] 0.8599088 0.7694664 0.9618755
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.8268985 5.3389389
#>   [2,] 2.1323066 5.0335308
#>   [3,] 2.1144190 5.0514184
#>   [4,] 1.4807975 5.6850399
#>   [5,] 2.0895275 5.0763099
#>   [6,] 2.7049091 4.4609283
#>   [7,] 2.2476404 4.9181970
#>   [8,] 0.0000000 7.1658374
#>   [9,] 2.2207508 4.9450866
#>  [10,] 2.4131599 4.7526775
#>  [11,] 0.0000000 7.1658374
#>  [12,] 0.7399534 6.4258840
#>  [13,] 2.2885881 4.8772493
#>  [14,] 1.3449525 5.8208849
#>  [15,] 2.4426730 4.7231644
#>  [16,] 0.0000000 7.1658374
#>  [17,] 2.0531858 5.1126516
#>  [18,] 1.3449525 5.8208849
#>  [19,] 2.3542287 4.8116087
#>  [20,] 1.3449525 5.8208849
#>  [21,] 1.6293753 5.5364621
#>  [22,] 1.9441819 5.2216555
#>  [23,] 2.9179252 4.2479122
#>  [24,] 2.0666660 5.0991714
#>  [25,] 0.7694664 6.3963709
#>  [26,] 0.6810222 6.4848152
#>  [27,] 0.7694664 6.3963709
#>  [28,] 1.8217843 5.3440531
#>  [29,] 1.8086585 5.3571789
#>  [30,] 1.3288003 5.8370370
#>  [31,] 0.7113310 6.4545064
#>  [32,] 0.4819460 6.6838914
#>  [33,] 0.4819460 6.6838914
#>  [34,] 1.3449525 5.8208849
#>  [35,] 0.9618755 6.2039619
#>  [36,] 2.1144190 5.0514184
#>  [37,] 2.9124181 4.2534193
#>  [38,] 2.0259747 5.1398627
#>  [39,] 0.0000000 7.1658374
#>  [40,] 0.8599088 6.3059286
#>  [41,] 1.2514124 5.9144250
#>  [42,] 1.2514124 5.9144250
#>  [43,] 1.3418548 5.8239826
#>  [44,] 2.1113212 5.0545162
#>  [45,] 0.4819460 6.6838914
#>  [46,] 0.4819460 6.6838914
#>  [47,] 0.4819460 6.6838914
#>  [48,] 0.0000000 7.1658374
#>  [49,] 2.5331153 4.6327221
#>  [50,] 0.7113310 6.4545064
#>  [51,] 1.5094198 5.6564176
#>  [52,] 1.2514124 5.9144250
#>  [53,] 0.7694664 6.3963709
#>  [54,] 1.1791203 5.9867171
#>  [55,] 1.2218993 5.9439380
#>  [56,] 2.1248014 5.0410360
#>  [57,] 1.8667940 5.2990434
#>  [58,] 1.8372809 5.3285565
#>  [59,] 1.9913658 5.1744716
#>  [60,] 1.8667940 5.2990434
#>  [61,] 2.7373057 4.4285317
#>  [62,] 2.0895275 5.0763099
#>  [63,] 2.5668519 4.5989855
#>  [64,] 0.6971743 6.4686630
#>  [65,] 2.1618196 5.0040178
#>  [66,] 1.8991906 5.2666468
#>  [67,] 1.8742991 5.2915382
#>  [68,] 0.6477782 6.5180592
#>  [69,] 2.1551525 5.0106849
#>  [70,] 1.1932770 5.9725604
#>  [71,] 4.8886840 2.2771534
#>  [72,] 5.0545162 2.1113212
#>  [73,] 5.2086011 1.9572363
#>  [74,] 5.6905470 1.4752904
#>  [75,] 5.3389389 1.8268985
#>  [76,] 5.9725604 1.1932770
#>  [77,] 4.8448527 2.3209847
#>  [78,] 4.2799539 2.8858835
#>  [79,] 6.0361133 1.1297241
#>  [80,] 5.3550911 1.8107463
#>  [81,] 5.7220159 1.4438215
#>  [82,] 5.3571789 1.8086585
#>  [83,] 5.5885803 1.5772571
#>  [84,] 6.2039619 0.9618755
#>  [85,] 5.9439380 1.2218993
#>  [86,] 5.5067875 1.6590499
#>  [87,] 5.1398627 2.0259747
#>  [88,] 6.3963709 0.7694664
#>  [89,] 6.5180592 0.6477782
#>  [90,] 6.0685099 1.0973275
#>  [91,] 5.0427689 2.1230685
#>  [92,] 5.9867171 1.1791203
#>  [93,] 4.9933727 2.1724647
#>  [94,] 4.7534732 2.4123642
#>  [95,] 5.0095248 2.1563126
#>  [96,] 5.7448618 1.4209756
#>  [97,] 5.7220159 1.4438215
#>  [98,] 4.7965088 2.3693286
#>  [99,] 5.7809894 1.3848480
#> [100,] 5.4640085 1.7018289
#> [101,] 6.3963709 0.7694664
#> [102,] 6.3059286 0.8599088
#> [103,] 5.8532815 1.3125559
#> [104,] 5.4926309 1.6732065
#> [105,] 4.8486269 2.3172105
#> [106,] 4.6962749 2.4695625
#> [107,] 4.4951823 2.6706551
#> [108,] 5.7448618 1.4209756
#> [109,] 4.1676047 2.9982327
#> [110,] 4.7534732 2.4123642
#> [111,] 6.6838914 0.4819460
#> [112,] 5.4640085 1.7018289
#> [113,] 4.9210806 2.2447568
#> [114,] 5.8694336 1.2964038
#> [115,] 4.6761567 2.4896807
#> [116,] 5.4207317 1.7451057
#> [117,] 5.6087542 1.5570831
#> [118,] 7.1658374 0.0000000
#> [119,] 5.6991966 1.4666408
#> [120,] 5.0970837 2.0687537
#> [121,] 4.8486269 2.3172105
#> [122,] 3.8346332 3.3312042
#> [123,] 6.2039619 0.9618755
#> [124,] 7.1658374 0.0000000
#> [125,] 7.1658374 0.0000000
#> [126,] 6.3963709 0.7694664
#> [127,] 6.2039619 0.9618755
#> [128,] 6.4258840 0.7399534
#> [129,] 5.4344954 1.7313420
#> [130,] 5.8532815 1.3125559
#> [131,] 6.5180592 0.6477782
#> [132,] 5.8105024 1.3553350
#> [133,] 5.0970837 2.0687537
#> [134,] 7.1658374 0.0000000
#> [135,] 7.1658374 0.0000000
#> [136,] 7.1658374 0.0000000
#> [137,] 7.1658374 0.0000000
#> [138,] 4.5114267 2.6544107
#> [139,] 4.3932578 2.7725796
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.25494557 0.74505443
#>   [2,] 0.29756558 0.70243442
#>   [3,] 0.29506935 0.70493065
#>   [4,] 0.20664681 0.79335319
#>   [5,] 0.29159572 0.70840428
#>   [6,] 0.37747286 0.62252714
#>   [7,] 0.31366053 0.68633947
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.30990807 0.69009193
#>  [10,] 0.33675896 0.66324104
#>  [11,] 0.00000000 1.00000000
#>  [12,] 0.10326126 0.89673874
#>  [13,] 0.31937483 0.68062517
#>  [14,] 0.18768951 0.81231049
#>  [15,] 0.34087753 0.65912247
#>  [16,] 0.00000000 1.00000000
#>  [17,] 0.28652419 0.71347581
#>  [18,] 0.18768951 0.81231049
#>  [19,] 0.32853505 0.67146495
#>  [20,] 0.18768951 0.81231049
#>  [21,] 0.22738100 0.77261900
#>  [22,] 0.27131259 0.72868741
#>  [23,] 0.40719947 0.59280053
#>  [24,] 0.28840537 0.71159463
#>  [25,] 0.10737984 0.89262016
#>  [26,] 0.09503735 0.90496265
#>  [27,] 0.10737984 0.89262016
#>  [28,] 0.25423188 0.74576812
#>  [29,] 0.25240016 0.74759984
#>  [30,] 0.18543546 0.81456454
#>  [31,] 0.09926698 0.90073302
#>  [32,] 0.06725606 0.93274394
#>  [33,] 0.06725606 0.93274394
#>  [34,] 0.18768951 0.81231049
#>  [35,] 0.13423072 0.86576928
#>  [36,] 0.29506935 0.70493065
#>  [37,] 0.40643095 0.59356905
#>  [38,] 0.28272686 0.71727314
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.12000116 0.87999884
#>  [41,] 0.17463589 0.82536411
#>  [42,] 0.17463589 0.82536411
#>  [43,] 0.18725722 0.81274278
#>  [44,] 0.29463705 0.70536295
#>  [45,] 0.06725606 0.93274394
#>  [46,] 0.06725606 0.93274394
#>  [47,] 0.06725606 0.93274394
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.35349886 0.64650114
#>  [50,] 0.09926698 0.90073302
#>  [51,] 0.21064109 0.78935891
#>  [52,] 0.17463589 0.82536411
#>  [53,] 0.10737984 0.89262016
#>  [54,] 0.16454745 0.83545255
#>  [55,] 0.17051731 0.82948269
#>  [56,] 0.29651823 0.70348177
#>  [57,] 0.26051302 0.73948698
#>  [58,] 0.25639445 0.74360555
#>  [59,] 0.27789715 0.72210285
#>  [60,] 0.26051302 0.73948698
#>  [61,] 0.38199383 0.61800617
#>  [62,] 0.29159572 0.70840428
#>  [63,] 0.35820682 0.64179318
#>  [64,] 0.09729140 0.90270860
#>  [65,] 0.30168416 0.69831584
#>  [66,] 0.26503400 0.73496600
#>  [67,] 0.26156038 0.73843962
#>  [68,] 0.09039811 0.90960189
#>  [69,] 0.30075375 0.69924625
#>  [70,] 0.16652303 0.83347697
#>  [71,] 0.68222089 0.31777911
#>  [72,] 0.70536295 0.29463705
#>  [73,] 0.72686565 0.27313435
#>  [74,] 0.79412171 0.20587829
#>  [75,] 0.74505443 0.25494557
#>  [76,] 0.83347697 0.16652303
#>  [77,] 0.67610419 0.32389581
#>  [78,] 0.59727198 0.40272802
#>  [79,] 0.84234583 0.15765417
#>  [80,] 0.74730849 0.25269151
#>  [81,] 0.79851322 0.20148678
#>  [82,] 0.74759984 0.25240016
#>  [83,] 0.77989214 0.22010786
#>  [84,] 0.86576928 0.13423072
#>  [85,] 0.82948269 0.17051731
#>  [86,] 0.76847788 0.23152212
#>  [87,] 0.71727314 0.28272686
#>  [88,] 0.89262016 0.10737984
#>  [89,] 0.90960189 0.09039811
#>  [90,] 0.84686681 0.15313319
#>  [91,] 0.70372360 0.29627640
#>  [92,] 0.83545255 0.16454745
#>  [93,] 0.69683031 0.30316969
#>  [94,] 0.66335209 0.33664791
#>  [95,] 0.69908436 0.30091564
#>  [96,] 0.80170139 0.19829861
#>  [97,] 0.79851322 0.20148678
#>  [98,] 0.66935774 0.33064226
#>  [99,] 0.80674303 0.19325697
#> [100,] 0.76250802 0.23749198
#> [101,] 0.89262016 0.10737984
#> [102,] 0.87999884 0.12000116
#> [103,] 0.81683147 0.18316853
#> [104,] 0.76650230 0.23349770
#> [105,] 0.67663089 0.32336911
#> [106,] 0.65537001 0.34462999
#> [107,] 0.62730732 0.37269268
#> [108,] 0.80170139 0.19829861
#> [109,] 0.58159354 0.41840646
#> [110,] 0.66335209 0.33664791
#> [111,] 0.93274394 0.06725606
#> [112,] 0.76250802 0.23749198
#> [113,] 0.68674187 0.31325813
#> [114,] 0.81908552 0.18091448
#> [115,] 0.65256249 0.34743751
#> [116,] 0.75646870 0.24353130
#> [117,] 0.78270744 0.21729256
#> [118,] 1.00000000 0.00000000
#> [119,] 0.79532876 0.20467124
#> [120,] 0.71130328 0.28869672
#> [121,] 0.67663089 0.32336911
#> [122,] 0.53512702 0.46487298
#> [123,] 0.86576928 0.13423072
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 0.89262016 0.10737984
#> [127,] 0.86576928 0.13423072
#> [128,] 0.89673874 0.10326126
#> [129,] 0.75838944 0.24161056
#> [130,] 0.81683147 0.18316853
#> [131,] 0.90960189 0.09039811
#> [132,] 0.81086161 0.18913839
#> [133,] 0.71130328 0.28869672
#> [134,] 1.00000000 0.00000000
#> [135,] 1.00000000 0.00000000
#> [136,] 1.00000000 0.00000000
#> [137,] 1.00000000 0.00000000
#> [138,] 0.62957425 0.37042575
#> [139,] 0.61308366 0.38691634
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.8883986  4.5477120  3.4630139 13.3979951  0.0000000  3.0837739  1.7896461 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.3600092  0.0000000  0.0000000  0.0000000  0.0000000  2.3701556  0.9356967 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  2.6427869  4.1306999  1.7764073  0.0000000  1.5257755  5.1472643  2.3516959 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  3.7760029  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.4703679  2.5685539  0.0000000  0.0000000  0.0000000  2.7817552  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  2.8990245  1.6764556  6.2492647  0.0000000  3.5253178 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  0.0000000  1.6122000  0.0000000  0.0000000  0.0000000  4.4326472  1.2505044 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  2.5849834  0.5475682  0.0000000  2.8799053  0.0000000  6.9425424  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.7590974  4.6327784 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 69 70 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12        V59        V45        V27         V9        V10        V52 
#> 13.3979951  6.9425424  6.2492647  5.1472643  4.6327784  4.5477120  4.4326472 
#>        V23        V31        V47        V11        V14        V43        V57 
#>  4.1306999  3.7760029  3.5253178  3.4630139  3.0837739  2.8990245  2.8799053 
#>         V4        V22        V54        V36        V20        V28        V15 
#>  2.7817552  2.6427869  2.5849834  2.5685539  2.3701556  2.3516959  1.7896461 
#>        V24        V44        V49        V26        V35        V16        V53 
#>  1.7764073  1.6764556  1.6122000  1.5257755  1.4703679  1.3600092  1.2505044 
#>        V21         V1         V8        V55        V13        V17        V18 
#>  0.9356967  0.8883986  0.7590974  0.5475682  0.0000000  0.0000000  0.0000000 
#>        V19         V2        V25        V29         V3        V30        V32 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V33        V34        V37        V38        V39        V40        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V42        V46        V48         V5        V50        V51        V56 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58         V6        V60         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1449275