Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V49>=0.04575 76 15 M (0.80263158 0.19736842)  
#>      4) V12>=0.1941 56  2 M (0.96428571 0.03571429) *
#>      5) V12< 0.1941 20  7 R (0.35000000 0.65000000)  
#>       10) V20>=0.4866 7  0 M (1.00000000 0.00000000) *
#>       11) V20< 0.4866 13  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.04575 63 11 R (0.17460317 0.82539683)  
#>      6) V31< 0.4072 21 10 R (0.47619048 0.52380952)  
#>       12) V60>=0.008 7  0 M (1.00000000 0.00000000) *
#>       13) V60< 0.008 14  3 R (0.21428571 0.78571429) *
#>      7) V31>=0.4072 42  1 R (0.02380952 0.97619048) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 56 M (0.59712230 0.40287770)  
#>    2) V11>=0.18875 92 23 M (0.75000000 0.25000000)  
#>      4) V27>=0.8191 45  0 M (1.00000000 0.00000000) *
#>      5) V27< 0.8191 47 23 M (0.51063830 0.48936170)  
#>       10) V56>=0.00625 29  6 M (0.79310345 0.20689655)  
#>         20) V21>=0.6543 22  1 M (0.95454545 0.04545455) *
#>         21) V21< 0.6543 7  2 R (0.28571429 0.71428571) *
#>       11) V56< 0.00625 18  1 R (0.05555556 0.94444444) *
#>    3) V11< 0.18875 47 14 R (0.29787234 0.70212766)  
#>      6) V16>=0.3645 16  3 M (0.81250000 0.18750000) *
#>      7) V16< 0.3645 31  1 R (0.03225806 0.96774194) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V11>=0.19615 91 20 M (0.78021978 0.21978022)  
#>      4) V51>=0.01285 59  2 M (0.96610169 0.03389831) *
#>      5) V51< 0.01285 32 14 R (0.43750000 0.56250000)  
#>       10) V33>=0.2395 18  4 M (0.77777778 0.22222222) *
#>       11) V33< 0.2395 14  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.19615 48  9 R (0.18750000 0.81250000)  
#>      6) V59>=0.01035 9  2 M (0.77777778 0.22222222) *
#>      7) V59< 0.01035 39  2 R (0.05128205 0.94871795) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V35< 0.2208 46  4 M (0.91304348 0.08695652) *
#>    3) V35>=0.2208 93 32 R (0.34408602 0.65591398)  
#>      6) V44>=0.3511 11  0 M (1.00000000 0.00000000) *
#>      7) V44< 0.3511 82 21 R (0.25609756 0.74390244)  
#>       14) V23>=0.7803 29 10 M (0.65517241 0.34482759)  
#>         28) V37< 0.4651 21  2 M (0.90476190 0.09523810) *
#>         29) V37>=0.4651 8  0 R (0.00000000 1.00000000) *
#>       15) V23< 0.7803 53  2 R (0.03773585 0.96226415) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V52>=0.01405 53 11 M (0.79245283 0.20754717)  
#>      4) V32< 0.49425 39  1 M (0.97435897 0.02564103) *
#>      5) V32>=0.49425 14  4 R (0.28571429 0.71428571) *
#>    3) V52< 0.01405 86 25 R (0.29069767 0.70930233)  
#>      6) V21>=0.65735 44 22 M (0.50000000 0.50000000)  
#>       12) V1>=0.0128 28  6 M (0.78571429 0.21428571)  
#>         24) V15< 0.5467 21  0 M (1.00000000 0.00000000) *
#>         25) V15>=0.5467 7  1 R (0.14285714 0.85714286) *
#>       13) V1< 0.0128 16  0 R (0.00000000 1.00000000) *
#>      7) V21< 0.65735 42  3 R (0.07142857 0.92857143) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V9>=0.1645 73 15 M (0.79452055 0.20547945)  
#>      4) V52>=0.0047 65  8 M (0.87692308 0.12307692) *
#>      5) V52< 0.0047 8  1 R (0.12500000 0.87500000) *
#>    3) V9< 0.1645 66 16 R (0.24242424 0.75757576)  
#>      6) V32< 0.3314 23  9 M (0.60869565 0.39130435)  
#>       12) V44>=0.1862 12  0 M (1.00000000 0.00000000) *
#>       13) V44< 0.1862 11  2 R (0.18181818 0.81818182) *
#>      7) V32>=0.3314 43  2 R (0.04651163 0.95348837) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V39>=0.1744 101 40 M (0.60396040 0.39603960)  
#>      4) V50>=0.01785 41  5 M (0.87804878 0.12195122)  
#>        8) V10>=0.09335 34  1 M (0.97058824 0.02941176) *
#>        9) V10< 0.09335 7  3 R (0.42857143 0.57142857) *
#>      5) V50< 0.01785 60 25 R (0.41666667 0.58333333)  
#>       10) V42>=0.29755 19  2 M (0.89473684 0.10526316) *
#>       11) V42< 0.29755 41  8 R (0.19512195 0.80487805) *
#>    3) V39< 0.1744 38  3 R (0.07894737 0.92105263) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V28>=0.9202 40  1 M (0.97500000 0.02500000) *
#>    3) V28< 0.9202 99 42 R (0.42424242 0.57575758)  
#>      6) V1>=0.02045 65 25 M (0.61538462 0.38461538)  
#>       12) V51>=0.01305 37  4 M (0.89189189 0.10810811)  
#>         24) V12>=0.1231 30  0 M (1.00000000 0.00000000) *
#>         25) V12< 0.1231 7  3 R (0.42857143 0.57142857) *
#>       13) V51< 0.01305 28  7 R (0.25000000 0.75000000)  
#>         26) V36< 0.44355 11  4 M (0.63636364 0.36363636) *
#>         27) V36>=0.44355 17  0 R (0.00000000 1.00000000) *
#>      7) V1< 0.02045 34  2 R (0.05882353 0.94117647) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 56 M (0.59712230 0.40287770)  
#>    2) V47>=0.06225 111 32 M (0.71171171 0.28828829)  
#>      4) V37< 0.48 80 11 M (0.86250000 0.13750000)  
#>        8) V15< 0.57285 71  4 M (0.94366197 0.05633803)  
#>         16) V9>=0.109 64  0 M (1.00000000 0.00000000) *
#>         17) V9< 0.109 7  3 R (0.42857143 0.57142857) *
#>        9) V15>=0.57285 9  2 R (0.22222222 0.77777778) *
#>      5) V37>=0.48 31 10 R (0.32258065 0.67741935)  
#>       10) V27>=0.82755 8  0 M (1.00000000 0.00000000) *
#>       11) V27< 0.82755 23  2 R (0.08695652 0.91304348) *
#>    3) V47< 0.06225 28  4 R (0.14285714 0.85714286)  
#>      6) V53< 0.0046 7  3 M (0.57142857 0.42857143) *
#>      7) V53>=0.0046 21  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.5107914 0.4892086)  
#>    2) V48>=0.0503 99 32 M (0.6767677 0.3232323)  
#>      4) V51>=0.01305 60  7 M (0.8833333 0.1166667) *
#>      5) V51< 0.01305 39 14 R (0.3589744 0.6410256)  
#>       10) V59>=0.0104 10  0 M (1.0000000 0.0000000) *
#>       11) V59< 0.0104 29  4 R (0.1379310 0.8620690)  
#>         22) V14< 0.25965 7  3 M (0.5714286 0.4285714) *
#>         23) V14>=0.25965 22  0 R (0.0000000 1.0000000) *
#>    3) V48< 0.0503 40  4 R (0.1000000 0.9000000) *
#> 
#> 
#> $weights
#>  [1] 0.8090480 0.8476408 0.7270150 0.7069392 0.7366673 0.6694236 0.5799773
#>  [8] 0.7974748 0.7966370 0.6146628
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.9015792 5.3839065
#>   [2,] 2.1738375 5.1116482
#>   [3,] 2.3145392 4.9709466
#>   [4,] 2.2314289 5.0540569
#>   [5,] 2.1317273 5.1537585
#>   [6,] 2.4417526 4.8437332
#>   [7,] 2.3431901 4.9422957
#>   [8,] 1.3216020 5.9638838
#>   [9,] 0.6146628 6.6708230
#>  [10,] 2.0582693 5.2272164
#>  [11,] 2.1893186 5.0961672
#>  [12,] 0.7270150 6.5584708
#>  [13,] 2.1391526 5.1463332
#>  [14,] 2.8758199 4.4096659
#>  [15,] 2.2242551 5.0612307
#>  [16,] 2.9000148 4.3854710
#>  [17,] 1.3964386 5.8890472
#>  [18,] 1.9216551 5.3638307
#>  [19,] 1.4121376 5.8733482
#>  [20,] 1.1946401 6.0908457
#>  [21,] 0.0000000 7.2854858
#>  [22,] 0.8476408 6.4378450
#>  [23,] 2.6731117 4.6123741
#>  [24,] 1.5843082 5.7011776
#>  [25,] 1.5843082 5.7011776
#>  [26,] 1.4636823 5.8218035
#>  [27,] 2.2151389 5.0703468
#>  [28,] 1.5035762 5.7819096
#>  [29,] 0.7974748 6.4880110
#>  [30,] 1.1946401 6.0908457
#>  [31,] 1.3964386 5.8890472
#>  [32,] 0.0000000 7.2854858
#>  [33,] 2.0486170 5.2368688
#>  [34,] 1.3069922 5.9784936
#>  [35,] 1.3216020 5.9638838
#>  [36,] 0.7069392 6.5785466
#>  [37,] 2.1893186 5.0961672
#>  [38,] 0.0000000 7.2854858
#>  [39,] 1.5035762 5.7819096
#>  [40,] 2.1190768 5.1664090
#>  [41,] 1.3216020 5.9638838
#>  [42,] 0.0000000 7.2854858
#>  [43,] 0.0000000 7.2854858
#>  [44,] 0.7366673 6.5488184
#>  [45,] 0.7974748 6.4880110
#>  [46,] 0.6146628 6.6708230
#>  [47,] 1.3416778 5.9438080
#>  [48,] 2.0036881 5.2817977
#>  [49,] 1.5941118 5.6913740
#>  [50,] 1.3766143 5.9088715
#>  [51,] 0.0000000 7.2854858
#>  [52,] 1.5545800 5.7309058
#>  [53,] 1.5545800 5.7309058
#>  [54,] 1.5545800 5.7309058
#>  [55,] 1.9921149 5.2933709
#>  [56,] 2.9124257 4.3730601
#>  [57,] 2.3721306 4.9133552
#>  [58,] 2.0111014 5.2743844
#>  [59,] 2.1331059 5.1523799
#>  [60,] 1.3416778 5.9438080
#>  [61,] 0.6694236 6.6160622
#>  [62,] 2.9011040 4.3843818
#>  [63,] 1.5333044 5.7521814
#>  [64,] 2.1479672 5.1375186
#>  [65,] 5.7309058 1.5545800
#>  [66,] 4.9709466 2.3145392
#>  [67,] 4.4652650 2.8202208
#>  [68,] 4.9342688 2.3512170
#>  [69,] 5.1124860 2.1729998
#>  [70,] 4.4854122 2.8000736
#>  [71,] 5.2019323 2.0835534
#>  [72,] 4.2129024 3.0725834
#>  [73,] 6.6708230 0.6146628
#>  [74,] 4.9643590 2.3211268
#>  [75,] 5.9088715 1.3766143
#>  [76,] 5.8194252 1.4660606
#>  [77,] 5.2817977 2.0036881
#>  [78,] 5.1923514 2.0931344
#>  [79,] 5.6403702 1.6451156
#>  [80,] 6.6160622 0.6694236
#>  [81,] 6.6708230 0.6146628
#>  [82,] 4.5000220 2.7854638
#>  [83,] 5.1463332 2.1391526
#>  [84,] 6.0360849 1.2494009
#>  [85,] 5.2817977 2.0036881
#>  [86,] 6.4764378 0.8090480
#>  [87,] 5.7521814 1.5333044
#>  [88,] 5.6412079 1.6442779
#>  [89,] 4.9844348 2.3010510
#>  [90,] 5.7819096 1.5035762
#>  [91,] 6.4888488 0.7966370
#>  [92,] 4.9342688 2.3512170
#>  [93,] 4.9431334 2.3423524
#>  [94,] 6.4378450 0.8476408
#>  [95,] 4.9717843 2.3137015
#>  [96,] 5.8218035 1.4636823
#>  [97,] 5.7011776 1.5843082
#>  [98,] 6.4378450 0.8476408
#>  [99,] 5.8890472 1.3964386
#> [100,] 4.5547827 2.7307030
#> [101,] 7.2854858 0.0000000
#> [102,] 5.8185874 1.4668984
#> [103,] 4.3719708 2.9135150
#> [104,] 6.5488184 0.7366673
#> [105,] 5.7513437 1.5341421
#> [106,] 6.5488184 0.7366673
#> [107,] 6.5488184 0.7366673
#> [108,] 6.4378450 0.8476408
#> [109,] 4.3941854 2.8913004
#> [110,] 5.8617750 1.4237108
#> [111,] 5.7810718 1.5044139
#> [112,] 5.7694986 1.5159871
#> [113,] 5.8185874 1.4668984
#> [114,] 5.2817977 2.0036881
#> [115,] 5.2817977 2.0036881
#> [116,] 3.7773838 3.5081020
#> [117,] 6.7055085 0.5799773
#> [118,] 6.5488184 0.7366673
#> [119,] 6.5488184 0.7366673
#> [120,] 7.2854858 0.0000000
#> [121,] 6.5488184 0.7366673
#> [122,] 5.0865148 2.1989710
#> [123,] 5.9688412 1.3166446
#> [124,] 5.2994176 1.9860682
#> [125,] 5.9688412 1.3166446
#> [126,] 5.0703468 2.2151389
#> [127,] 5.9688412 1.3166446
#> [128,] 5.9688412 1.3166446
#> [129,] 6.7055085 0.5799773
#> [130,] 6.7055085 0.5799773
#> [131,] 6.7055085 0.5799773
#> [132,] 6.7055085 0.5799773
#> [133,] 5.8964605 1.3890252
#> [134,] 6.7055085 0.5799773
#> [135,] 6.7055085 0.5799773
#> [136,] 6.5488184 0.7366673
#> [137,] 5.9688412 1.3166446
#> [138,] 5.9688412 1.3166446
#> [139,] 5.2817977 2.0036881
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.26100926 0.73899074
#>   [2,] 0.29837922 0.70162078
#>   [3,] 0.31769182 0.68230818
#>   [4,] 0.30628416 0.69371584
#>   [5,] 0.29259919 0.70740081
#>   [6,] 0.33515303 0.66484697
#>   [7,] 0.32162442 0.67837558
#>   [8,] 0.18140204 0.81859796
#>   [9,] 0.08436813 0.91563187
#>  [10,] 0.28251642 0.71748358
#>  [11,] 0.30050414 0.69949586
#>  [12,] 0.09978950 0.90021050
#>  [13,] 0.29361839 0.70638161
#>  [14,] 0.39473276 0.60526724
#>  [15,] 0.30529949 0.69470051
#>  [16,] 0.39805373 0.60194627
#>  [17,] 0.19167405 0.80832595
#>  [18,] 0.26376485 0.73623515
#>  [19,] 0.19382889 0.80617111
#>  [20,] 0.16397535 0.83602465
#>  [21,] 0.00000000 1.00000000
#>  [22,] 0.11634651 0.88365349
#>  [23,] 0.36690919 0.63309081
#>  [24,] 0.21746088 0.78253912
#>  [25,] 0.21746088 0.78253912
#>  [26,] 0.20090387 0.79909613
#>  [27,] 0.30404821 0.69595179
#>  [28,] 0.20637968 0.79362032
#>  [29,] 0.10946076 0.89053924
#>  [30,] 0.16397535 0.83602465
#>  [31,] 0.19167405 0.80832595
#>  [32,] 0.00000000 1.00000000
#>  [33,] 0.28119154 0.71880846
#>  [34,] 0.17939672 0.82060328
#>  [35,] 0.18140204 0.81859796
#>  [36,] 0.09703391 0.90296609
#>  [37,] 0.30050414 0.69949586
#>  [38,] 0.00000000 1.00000000
#>  [39,] 0.20637968 0.79362032
#>  [40,] 0.29086280 0.70913720
#>  [41,] 0.18140204 0.81859796
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.10111438 0.89888562
#>  [45,] 0.10946076 0.89053924
#>  [46,] 0.08436813 0.91563187
#>  [47,] 0.18415763 0.81584237
#>  [48,] 0.27502463 0.72497537
#>  [49,] 0.21880652 0.78119348
#>  [50,] 0.18895298 0.81104702
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.21338042 0.78661958
#>  [53,] 0.21338042 0.78661958
#>  [54,] 0.21338042 0.78661958
#>  [55,] 0.27343611 0.72656389
#>  [56,] 0.39975724 0.60024276
#>  [57,] 0.32559676 0.67440324
#>  [58,] 0.27604218 0.72395782
#>  [59,] 0.29278843 0.70721157
#>  [60,] 0.18415763 0.81584237
#>  [61,] 0.09188455 0.90811545
#>  [62,] 0.39820324 0.60179676
#>  [63,] 0.21046014 0.78953986
#>  [64,] 0.29482827 0.70517173
#>  [65,] 0.78661958 0.21338042
#>  [66,] 0.68230818 0.31769182
#>  [67,] 0.61289873 0.38710127
#>  [68,] 0.67727382 0.32272618
#>  [69,] 0.70173577 0.29826423
#>  [70,] 0.61566412 0.38433588
#>  [71,] 0.71401311 0.28598689
#>  [72,] 0.57825964 0.42174036
#>  [73,] 0.91563187 0.08436813
#>  [74,] 0.68140398 0.31859602
#>  [75,] 0.81104702 0.18895298
#>  [76,] 0.79876968 0.20123032
#>  [77,] 0.72497537 0.27502463
#>  [78,] 0.71269803 0.28730197
#>  [79,] 0.77419274 0.22580726
#>  [80,] 0.90811545 0.09188455
#>  [81,] 0.91563187 0.08436813
#>  [82,] 0.61766945 0.38233055
#>  [83,] 0.70638161 0.29361839
#>  [84,] 0.82850823 0.17149177
#>  [85,] 0.72497537 0.27502463
#>  [86,] 0.88895071 0.11104929
#>  [87,] 0.78953986 0.21046014
#>  [88,] 0.77430773 0.22569227
#>  [89,] 0.68415957 0.31584043
#>  [90,] 0.79362032 0.20637968
#>  [91,] 0.89065423 0.10934577
#>  [92,] 0.67727382 0.32272618
#>  [93,] 0.67849057 0.32150943
#>  [94,] 0.88365349 0.11634651
#>  [95,] 0.68242317 0.31757683
#>  [96,] 0.79909613 0.20090387
#>  [97,] 0.78253912 0.21746088
#>  [98,] 0.88365349 0.11634651
#>  [99,] 0.80832595 0.19167405
#> [100,] 0.62518587 0.37481413
#> [101,] 1.00000000 0.00000000
#> [102,] 0.79865469 0.20134531
#> [103,] 0.60009324 0.39990676
#> [104,] 0.89888562 0.10111438
#> [105,] 0.78942487 0.21057513
#> [106,] 0.89888562 0.10111438
#> [107,] 0.89888562 0.10111438
#> [108,] 0.88365349 0.11634651
#> [109,] 0.60314240 0.39685760
#> [110,] 0.80458258 0.19541742
#> [111,] 0.79350534 0.20649466
#> [112,] 0.79191681 0.20808319
#> [113,] 0.79865469 0.20134531
#> [114,] 0.72497537 0.27502463
#> [115,] 0.72497537 0.27502463
#> [116,] 0.51848070 0.48151930
#> [117,] 0.92039278 0.07960722
#> [118,] 0.89888562 0.10111438
#> [119,] 0.89888562 0.10111438
#> [120,] 1.00000000 0.00000000
#> [121,] 0.89888562 0.10111438
#> [122,] 0.69817098 0.30182902
#> [123,] 0.81927841 0.18072159
#> [124,] 0.72739385 0.27260615
#> [125,] 0.81927841 0.18072159
#> [126,] 0.69595179 0.30404821
#> [127,] 0.81927841 0.18072159
#> [128,] 0.81927841 0.18072159
#> [129,] 0.92039278 0.07960722
#> [130,] 0.92039278 0.07960722
#> [131,] 0.92039278 0.07960722
#> [132,] 0.92039278 0.07960722
#> [133,] 0.80934350 0.19065650
#> [134,] 0.92039278 0.07960722
#> [135,] 0.92039278 0.07960722
#> [136,] 0.89888562 0.10111438
#> [137,] 0.81927841 0.18072159
#> [138,] 0.81927841 0.18072159
#> [139,] 0.72497537 0.27502463
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 5.3776257 0.5242910 7.1117895 3.1682893 0.0000000 0.5649714 3.2617455 2.8866699 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 0.0000000 0.0000000 1.9513202 2.6085774 0.0000000 2.6777474 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 4.5631442 3.6530966 0.0000000 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 1.2287225 3.5883069 1.8358950 0.0000000 3.7334511 1.1432839 4.5274329 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 2.3398291 0.0000000 0.0000000 0.0000000 1.9537011 0.0000000 3.3746110 0.0000000 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 3.0554280 3.0878914 5.8265979 0.0000000 1.5940027 7.1276470 4.6528495 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.7239133 0.0000000 0.0000000 2.7146123 0.0000000 0.0000000 3.2878842 0.0000000 
#>       V60        V7        V8        V9 
#> 1.2355298 0.0000000 0.0000000 4.6191427 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 75 64 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V51       V11       V49        V1       V52        V9       V27       V37 
#> 7.1276470 7.1117895 5.8265979 5.3776257 4.6528495 4.6191427 4.5631442 4.5274329 
#>       V35       V28       V32       V44       V59       V15       V12       V48 
#> 3.7334511 3.6530966 3.5883069 3.3746110 3.2878842 3.2617455 3.1682893 3.0878914 
#>       V47       V16       V56       V23       V21       V39       V42       V20 
#> 3.0554280 2.8866699 2.7146123 2.6777474 2.6085774 2.3398291 1.9537011 1.9513202 
#>       V33       V50       V60       V31       V36       V53       V14       V10 
#> 1.8358950 1.5940027 1.2355298 1.2287225 1.1432839 0.7239133 0.5649714 0.5242910 
#>       V13       V17       V18       V19        V2       V22       V24       V25 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V26       V29        V3       V30       V34       V38        V4       V40 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V41       V43       V45       V46        V5       V54       V55       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V58        V6        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986