Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.15805 97 28 M (0.71134021 0.28865979)  
#>      4) V17< 0.6741 76 13 M (0.82894737 0.17105263)  
#>        8) V54>=0.00755 56  3 M (0.94642857 0.05357143) *
#>        9) V54< 0.00755 20 10 M (0.50000000 0.50000000)  
#>         18) V2< 0.04935 12  2 M (0.83333333 0.16666667) *
#>         19) V2>=0.04935 8  0 R (0.00000000 1.00000000) *
#>      5) V17>=0.6741 21  6 R (0.28571429 0.71428571)  
#>       10) V39>=0.31575 8  2 M (0.75000000 0.25000000) *
#>       11) V39< 0.31575 13  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.15805 42  5 R (0.11904762 0.88095238)  
#>      6) V10>=0.16975 7  2 M (0.71428571 0.28571429) *
#>      7) V10< 0.16975 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.5395683 0.4604317)  
#>    2) V36< 0.44265 84 20 M (0.7619048 0.2380952)  
#>      4) V44>=0.17095 47  1 M (0.9787234 0.0212766) *
#>      5) V44< 0.17095 37 18 R (0.4864865 0.5135135)  
#>       10) V4>=0.04425 17  3 M (0.8235294 0.1764706) *
#>       11) V4< 0.04425 20  4 R (0.2000000 0.8000000)  
#>         22) V52>=0.01265 7  3 M (0.5714286 0.4285714) *
#>         23) V52< 0.01265 13  0 R (0.0000000 1.0000000) *
#>    3) V36>=0.44265 55 11 R (0.2000000 0.8000000)  
#>      6) V45>=0.19425 16  5 M (0.6875000 0.3125000) *
#>      7) V45< 0.19425 39  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V13>=0.2476 76 23 M (0.69736842 0.30263158)  
#>      4) V47>=0.02235 67 14 M (0.79104478 0.20895522)  
#>        8) V51>=0.013 41  0 M (1.00000000 0.00000000) *
#>        9) V51< 0.013 26 12 R (0.46153846 0.53846154)  
#>         18) V38>=0.22315 12  1 M (0.91666667 0.08333333) *
#>         19) V38< 0.22315 14  1 R (0.07142857 0.92857143) *
#>      5) V47< 0.02235 9  0 R (0.00000000 1.00000000) *
#>    3) V13< 0.2476 63 10 R (0.15873016 0.84126984)  
#>      6) V48>=0.07185 18  8 M (0.55555556 0.44444444) *
#>      7) V48< 0.07185 45  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 56 R (0.40287770 0.59712230)  
#>   2) V51>=0.01885 47 10 M (0.78723404 0.21276596)  
#>     4) V18>=0.17695 39  2 M (0.94871795 0.05128205) *
#>     5) V18< 0.17695 8  0 R (0.00000000 1.00000000) *
#>   3) V51< 0.01885 92 19 R (0.20652174 0.79347826)  
#>     6) V26>=0.9857 11  3 M (0.72727273 0.27272727) *
#>     7) V26< 0.9857 81 11 R (0.13580247 0.86419753) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V21>=0.5482 86 29 M (0.66279070 0.33720930)  
#>      4) V51>=0.01235 52  7 M (0.86538462 0.13461538)  
#>        8) V11>=0.11205 44  1 M (0.97727273 0.02272727) *
#>        9) V11< 0.11205 8  2 R (0.25000000 0.75000000) *
#>      5) V51< 0.01235 34 12 R (0.35294118 0.64705882)  
#>       10) V23>=0.8259 12  3 M (0.75000000 0.25000000) *
#>       11) V23< 0.8259 22  3 R (0.13636364 0.86363636) *
#>    3) V21< 0.5482 53  6 R (0.11320755 0.88679245)  
#>      6) V36< 0.21375 11  5 M (0.54545455 0.45454545) *
#>      7) V36>=0.21375 42  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V11>=0.1779 81 23 M (0.71604938 0.28395062)  
#>      4) V27>=0.79205 38  0 M (1.00000000 0.00000000) *
#>      5) V27< 0.79205 43 20 R (0.46511628 0.53488372)  
#>       10) V46>=0.29325 11  0 M (1.00000000 0.00000000) *
#>       11) V46< 0.29325 32  9 R (0.28125000 0.71875000)  
#>         22) V14>=0.53545 7  1 M (0.85714286 0.14285714) *
#>         23) V14< 0.53545 25  3 R (0.12000000 0.88000000) *
#>    3) V11< 0.1779 58 15 R (0.25862069 0.74137931)  
#>      6) V45>=0.16285 20  7 M (0.65000000 0.35000000)  
#>       12) V14>=0.13545 13  0 M (1.00000000 0.00000000) *
#>       13) V14< 0.13545 7  0 R (0.00000000 1.00000000) *
#>      7) V45< 0.16285 38  2 R (0.05263158 0.94736842) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 R (0.48920863 0.51079137)  
#>   2) V11>=0.1579 92 30 M (0.67391304 0.32608696)  
#>     4) V47>=0.03955 73 13 M (0.82191781 0.17808219)  
#>       8) V17< 0.703 59  4 M (0.93220339 0.06779661) *
#>       9) V17>=0.703 14  5 R (0.35714286 0.64285714) *
#>     5) V47< 0.03955 19  2 R (0.10526316 0.89473684) *
#>   3) V11< 0.1579 47  6 R (0.12765957 0.87234043)  
#>     6) V7>=0.13985 7  1 M (0.85714286 0.14285714) *
#>     7) V7< 0.13985 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V51>=0.0171 41  3 M (0.92682927 0.07317073) *
#>    3) V51< 0.0171 98 36 R (0.36734694 0.63265306)  
#>      6) V19>=0.91985 12  0 M (1.00000000 0.00000000) *
#>      7) V19< 0.91985 86 24 R (0.27906977 0.72093023)  
#>       14) V10>=0.16625 39 17 M (0.56410256 0.43589744)  
#>         28) V57< 0.0071 26  4 M (0.84615385 0.15384615)  
#>           56) V60>=0.00305 19  0 M (1.00000000 0.00000000) *
#>           57) V60< 0.00305 7  3 R (0.42857143 0.57142857) *
#>         29) V57>=0.0071 13  0 R (0.00000000 1.00000000) *
#>       15) V10< 0.16625 47  2 R (0.04255319 0.95744681) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V36< 0.44265 95 30 M (0.68421053 0.31578947)  
#>      4) V27< 0.54005 38  1 M (0.97368421 0.02631579) *
#>      5) V27>=0.54005 57 28 R (0.49122807 0.50877193)  
#>       10) V44>=0.1866 18  0 M (1.00000000 0.00000000) *
#>       11) V44< 0.1866 39 10 R (0.25641026 0.74358974)  
#>         22) V57< 0.0056 17  8 M (0.52941176 0.47058824) *
#>         23) V57>=0.0056 22  1 R (0.04545455 0.95454545) *
#>    3) V36>=0.44265 44  7 R (0.15909091 0.84090909)  
#>      6) V21>=0.82165 7  2 M (0.71428571 0.28571429) *
#>      7) V21< 0.82165 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V45>=0.1716 52  8 M (0.84615385 0.15384615)  
#>      4) V5>=0.04305 45  3 M (0.93333333 0.06666667) *
#>      5) V5< 0.04305 7  2 R (0.28571429 0.71428571) *
#>    3) V45< 0.1716 87 20 R (0.22988506 0.77011494)  
#>      6) V54>=0.0208 8  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.0208 79 12 R (0.15189873 0.84810127)  
#>       14) V35< 0.12675 12  5 M (0.58333333 0.41666667) *
#>       15) V35>=0.12675 67  5 R (0.07462687 0.92537313) *
#> 
#> 
#> $weights
#>  [1] 0.8630811 0.7599718 0.6906816 0.5761275 0.7830205 0.9120656 0.6879482
#>  [8] 0.7317764 0.7798002 0.6963182
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.4518376 5.0289534
#>   [2,] 0.6906816 6.7901094
#>   [3,] 2.9870136 4.4937775
#>   [4,] 1.4506534 6.0301376
#>   [5,] 0.0000000 7.4807910
#>   [6,] 1.9985855 5.4822056
#>   [7,] 0.6879482 6.7928429
#>   [8,] 3.1023916 4.3783994
#>   [9,] 2.9470870 4.5337040
#>  [10,] 2.1104061 5.3703849
#>  [11,] 1.4506534 6.0301376
#>  [12,] 0.6906816 6.7901094
#>  [13,] 0.7599718 6.7208192
#>  [14,] 2.1584299 5.3223611
#>  [15,] 1.5397720 5.9410190
#>  [16,] 2.4630949 5.0176961
#>  [17,] 1.4917482 5.9890429
#>  [18,] 3.4747038 4.0060872
#>  [19,] 1.9958520 5.4849390
#>  [20,] 2.2417108 5.2390802
#>  [21,] 3.0427774 4.4380136
#>  [22,] 0.0000000 7.4807910
#>  [23,] 1.5397720 5.9410190
#>  [24,] 2.1104061 5.3703849
#>  [25,] 1.5948575 5.8859336
#>  [26,] 2.3308294 5.1499616
#>  [27,] 1.6428813 5.8379098
#>  [28,] 2.2336739 5.2471171
#>  [29,] 0.6906816 6.7901094
#>  [30,] 1.3786298 6.1021613
#>  [31,] 2.2304536 5.2503375
#>  [32,] 0.7830205 6.6977705
#>  [33,] 3.0215110 4.4592800
#>  [34,] 0.0000000 7.4807910
#>  [35,] 0.0000000 7.4807910
#>  [36,] 1.3591480 6.1216430
#>  [37,] 1.6461016 5.8346894
#>  [38,] 1.9985855 5.4822056
#>  [39,] 0.6906816 6.7901094
#>  [40,] 0.6906816 6.7901094
#>  [41,] 0.7599718 6.7208192
#>  [42,] 1.5628207 5.9179703
#>  [43,] 1.3559277 6.1248633
#>  [44,] 2.1389482 5.3418428
#>  [45,] 1.5397720 5.9410190
#>  [46,] 0.0000000 7.4807910
#>  [47,] 2.2715484 5.2092427
#>  [48,] 0.7798002 6.7009909
#>  [49,] 1.5429923 5.9377987
#>  [50,] 1.3842663 6.0965247
#>  [51,] 2.2360901 5.2447009
#>  [52,] 1.4562900 6.0245011
#>  [53,] 0.7317764 6.7490147
#>  [54,] 1.5115765 5.9692145
#>  [55,] 0.7317764 6.7490147
#>  [56,] 1.4506534 6.0301376
#>  [57,] 1.4506534 6.0301376
#>  [58,] 1.9985855 5.4822056
#>  [59,] 1.5628207 5.9179703
#>  [60,] 3.3181391 4.1626520
#>  [61,] 2.2309405 5.2498505
#>  [62,] 0.0000000 7.4807910
#>  [63,] 5.9179703 1.5628207
#>  [64,] 5.2608215 2.2199695
#>  [65,] 5.2334436 2.2473474
#>  [66,] 5.1844591 2.2963320
#>  [67,] 5.4567435 2.0240475
#>  [68,] 4.6083316 2.8724595
#>  [69,] 5.3098061 2.1709850
#>  [70,] 6.2139819 1.2668091
#>  [71,] 5.0093261 2.4714649
#>  [72,] 4.6218579 2.8589331
#>  [73,] 6.7208192 0.7599718
#>  [74,] 5.2962797 2.1845113
#>  [75,] 5.0176961 2.4630949
#>  [76,] 6.2083454 1.2724457
#>  [77,] 5.9410190 1.5397720
#>  [78,] 5.4849390 1.9958520
#>  [79,] 5.9297618 1.5510293
#>  [80,] 5.3930870 2.0877040
#>  [81,] 5.9297618 1.5510293
#>  [82,] 6.7844729 0.6963182
#>  [83,] 5.9297618 1.5510293
#>  [84,] 6.2167154 1.2640757
#>  [85,] 5.9297618 1.5510293
#>  [86,] 6.7844729 0.6963182
#>  [87,] 7.4807910 0.0000000
#>  [88,] 7.4807910 0.0000000
#>  [89,] 5.4341818 2.0466093
#>  [90,] 5.2962797 2.1845113
#>  [91,] 6.5687254 0.9120656
#>  [92,] 5.7857049 1.6950862
#>  [93,] 6.7844729 0.6963182
#>  [94,] 6.6977705 0.7830205
#>  [95,] 7.4807910 0.0000000
#>  [96,] 5.4285452 2.0522458
#>  [97,] 6.1216430 1.3591480
#>  [98,] 5.9179703 1.5628207
#>  [99,] 6.6977705 0.7830205
#> [100,] 7.4807910 0.0000000
#> [101,] 6.9046635 0.5761275
#> [102,] 4.5701399 2.9106511
#> [103,] 4.5132592 2.9675318
#> [104,] 6.2083454 1.2724457
#> [105,] 5.2585619 2.2222291
#> [106,] 4.6100665 2.8707246
#> [107,] 5.9179703 1.5628207
#> [108,] 5.3418428 2.1389482
#> [109,] 6.6977705 0.7830205
#> [110,] 5.2962797 2.1845113
#> [111,] 5.2608215 2.2199695
#> [112,] 6.2167154 1.2640757
#> [113,] 5.2785329 2.2022581
#> [114,] 4.6935485 2.7872425
#> [115,] 6.0937913 1.3869997
#> [116,] 6.1728872 1.3079039
#> [117,] 7.4807910 0.0000000
#> [118,] 5.2447009 2.2360901
#> [119,] 4.6055981 2.8751929
#> [120,] 7.4807910 0.0000000
#> [121,] 7.4807910 0.0000000
#> [122,] 7.4807910 0.0000000
#> [123,] 7.4807910 0.0000000
#> [124,] 6.6177099 0.8630811
#> [125,] 6.2139819 1.2668091
#> [126,] 6.9046635 0.5761275
#> [127,] 6.0245011 1.4562900
#> [128,] 7.4807910 0.0000000
#> [129,] 6.0046727 1.4761183
#> [130,] 6.2083454 1.2724457
#> [131,] 6.7844729 0.6963182
#> [132,] 7.4807910 0.0000000
#> [133,] 6.1728872 1.3079039
#> [134,] 6.9046635 0.5761275
#> [135,] 6.7844729 0.6963182
#> [136,] 6.2139819 1.2668091
#> [137,] 6.9046635 0.5761275
#> [138,] 6.7901094 0.6906816
#> [139,] 6.6977705 0.7830205
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32775112 0.67224888
#>   [2,] 0.09232735 0.90767265
#>   [3,] 0.39929114 0.60070886
#>   [4,] 0.19391711 0.80608289
#>   [5,] 0.00000000 1.00000000
#>   [6,] 0.26716232 0.73283768
#>   [7,] 0.09196195 0.90803805
#>   [8,] 0.41471438 0.58528562
#>   [9,] 0.39395393 0.60604607
#>  [10,] 0.28211002 0.71788998
#>  [11,] 0.19391711 0.80608289
#>  [12,] 0.09232735 0.90767265
#>  [13,] 0.10158977 0.89841023
#>  [14,] 0.28852964 0.71147036
#>  [15,] 0.20583010 0.79416990
#>  [16,] 0.32925594 0.67074406
#>  [17,] 0.19941049 0.80058951
#>  [18,] 0.46448347 0.53551653
#>  [19,] 0.26679692 0.73320308
#>  [20,] 0.29966227 0.70033773
#>  [21,] 0.40674541 0.59325459
#>  [22,] 0.00000000 1.00000000
#>  [23,] 0.20583010 0.79416990
#>  [24,] 0.28211002 0.71788998
#>  [25,] 0.21319369 0.78680631
#>  [26,] 0.31157526 0.68842474
#>  [27,] 0.21961331 0.78038669
#>  [28,] 0.29858793 0.70141207
#>  [29,] 0.09232735 0.90767265
#>  [30,] 0.18428930 0.81571070
#>  [31,] 0.29815745 0.70184255
#>  [32,] 0.10467082 0.89532918
#>  [33,] 0.40390261 0.59609739
#>  [34,] 0.00000000 1.00000000
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.18168507 0.81831493
#>  [37,] 0.22004379 0.77995621
#>  [38,] 0.26716232 0.73283768
#>  [39,] 0.09232735 0.90767265
#>  [40,] 0.09232735 0.90767265
#>  [41,] 0.10158977 0.89841023
#>  [42,] 0.20891116 0.79108884
#>  [43,] 0.18125459 0.81874541
#>  [44,] 0.28592541 0.71407459
#>  [45,] 0.20583010 0.79416990
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.30365082 0.69634918
#>  [48,] 0.10424034 0.89575966
#>  [49,] 0.20626059 0.79373941
#>  [50,] 0.18504277 0.81495723
#>  [51,] 0.29891092 0.70108908
#>  [52,] 0.19467059 0.80532941
#>  [53,] 0.09782072 0.90217928
#>  [54,] 0.20206106 0.79793894
#>  [55,] 0.09782072 0.90217928
#>  [56,] 0.19391711 0.80608289
#>  [57,] 0.19391711 0.80608289
#>  [58,] 0.26716232 0.73283768
#>  [59,] 0.20891116 0.79108884
#>  [60,] 0.44355457 0.55644543
#>  [61,] 0.29822254 0.70177746
#>  [62,] 0.00000000 1.00000000
#>  [63,] 0.79108884 0.20891116
#>  [64,] 0.70324401 0.29675599
#>  [65,] 0.69958426 0.30041574
#>  [66,] 0.69303621 0.30696379
#>  [67,] 0.72943403 0.27056597
#>  [68,] 0.61602196 0.38397804
#>  [69,] 0.70979206 0.29020794
#>  [70,] 0.83065840 0.16934160
#>  [71,] 0.66962519 0.33037481
#>  [72,] 0.61783010 0.38216990
#>  [73,] 0.89841023 0.10158977
#>  [74,] 0.70798392 0.29201608
#>  [75,] 0.67074406 0.32925594
#>  [76,] 0.82990493 0.17009507
#>  [77,] 0.79416990 0.20583010
#>  [78,] 0.73320308 0.26679692
#>  [79,] 0.79266507 0.20733493
#>  [80,] 0.72092469 0.27907531
#>  [81,] 0.79266507 0.20733493
#>  [82,] 0.90691918 0.09308082
#>  [83,] 0.79266507 0.20733493
#>  [84,] 0.83102380 0.16897620
#>  [85,] 0.79266507 0.20733493
#>  [86,] 0.90691918 0.09308082
#>  [87,] 1.00000000 0.00000000
#>  [88,] 1.00000000 0.00000000
#>  [89,] 0.72641807 0.27358193
#>  [90,] 0.70798392 0.29201608
#>  [91,] 0.87807898 0.12192102
#>  [92,] 0.77340816 0.22659184
#>  [93,] 0.90691918 0.09308082
#>  [94,] 0.89532918 0.10467082
#>  [95,] 1.00000000 0.00000000
#>  [96,] 0.72566460 0.27433540
#>  [97,] 0.81831493 0.18168507
#>  [98,] 0.79108884 0.20891116
#>  [99,] 0.89532918 0.10467082
#> [100,] 1.00000000 0.00000000
#> [101,] 0.92298575 0.07701425
#> [102,] 0.61091667 0.38908333
#> [103,] 0.60331310 0.39668690
#> [104,] 0.82990493 0.17009507
#> [105,] 0.70294196 0.29705804
#> [106,] 0.61625387 0.38374613
#> [107,] 0.79108884 0.20891116
#> [108,] 0.71407459 0.28592541
#> [109,] 0.89532918 0.10467082
#> [110,] 0.70798392 0.29201608
#> [111,] 0.70324401 0.29675599
#> [112,] 0.83102380 0.16897620
#> [113,] 0.70561160 0.29438840
#> [114,] 0.62741339 0.37258661
#> [115,] 0.81459183 0.18540817
#> [116,] 0.82516503 0.17483497
#> [117,] 1.00000000 0.00000000
#> [118,] 0.70108908 0.29891092
#> [119,] 0.61565657 0.38434343
#> [120,] 1.00000000 0.00000000
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 1.00000000 0.00000000
#> [124,] 0.88462703 0.11537297
#> [125,] 0.83065840 0.16934160
#> [126,] 0.92298575 0.07701425
#> [127,] 0.80532941 0.19467059
#> [128,] 1.00000000 0.00000000
#> [129,] 0.80267885 0.19732115
#> [130,] 0.82990493 0.17009507
#> [131,] 0.90691918 0.09308082
#> [132,] 1.00000000 0.00000000
#> [133,] 0.82516503 0.17483497
#> [134,] 0.92298575 0.07701425
#> [135,] 0.90691918 0.09308082
#> [136,] 0.83065840 0.16934160
#> [137,] 0.92298575 0.07701425
#> [138,] 0.90767265 0.09232735
#> [139,] 0.89532918 0.10467082
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  3.5238777 12.6823556  0.0000000  3.5711009  3.5491736  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  3.4998558  1.7809022  2.0720917  1.4884041  0.0000000  5.0474430 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  1.1844339  0.0000000  0.0000000  1.0098685  4.8643811  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.9487716  8.5221597  0.0000000  1.6495384  1.2438805  1.4046854  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  4.7195054  8.7677586  1.9954760  4.5299286 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.4179726  0.0000000  0.9152604  0.0000000 10.3877579  0.5841485  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.1937576  0.0000000  0.0000000  3.2553369  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.6323697  1.5578040  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 77 62 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V51        V45        V36        V21        V27        V44 
#> 12.6823556 10.3877579  8.7677586  8.5221597  5.0474430  4.8643811  4.7195054 
#>        V47        V13        V14        V10        V17        V57        V54 
#>  4.5299286  3.5711009  3.5491736  3.5238777  3.4998558  3.2553369  3.1937576 
#>        V19        V46        V18        V38         V7         V2        V48 
#>  2.0720917  1.9954760  1.7809022  1.6495384  1.5578040  1.4884041  1.4179726 
#>         V4        V39        V23        V26        V35         V5        V60 
#>  1.4046854  1.2438805  1.1844339  1.0098685  0.9487716  0.9152604  0.6323697 
#>        V52         V1        V12        V15        V16        V20        V22 
#>  0.5841485  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V25        V28        V29         V3        V30        V31 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V32        V33        V34        V37        V40        V41        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V43        V49        V50        V53        V55        V56        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3478261