Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 50 M (0.64028777 0.35971223)  
#>    2) V11>=0.21705 76 12 M (0.84210526 0.15789474)  
#>      4) V16< 0.6707 60  2 M (0.96666667 0.03333333) *
#>      5) V16>=0.6707 16  6 R (0.37500000 0.62500000) *
#>    3) V11< 0.21705 63 25 R (0.39682540 0.60317460)  
#>      6) V26< 0.59025 17  2 M (0.88235294 0.11764706) *
#>      7) V26>=0.59025 46 10 R (0.21739130 0.78260870)  
#>       14) V45>=0.16015 10  3 M (0.70000000 0.30000000) *
#>       15) V45< 0.16015 36  3 R (0.08333333 0.91666667) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V9>=0.14595 65 12 M (0.81538462 0.18461538)  
#>      4) V7< 0.2115 56  5 M (0.91071429 0.08928571) *
#>      5) V7>=0.2115 9  2 R (0.22222222 0.77777778) *
#>    3) V9< 0.14595 74 22 R (0.29729730 0.70270270)  
#>      6) V32< 0.1885 11  1 M (0.90909091 0.09090909) *
#>      7) V32>=0.1885 63 12 R (0.19047619 0.80952381)  
#>       14) V46>=0.16605 10  3 M (0.70000000 0.30000000) *
#>       15) V46< 0.16605 53  5 R (0.09433962 0.90566038) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V11>=0.15805 103 24 M (0.76699029 0.23300971)  
#>      4) V39>=0.10345 94 16 M (0.82978723 0.17021277)  
#>        8) V53>=0.00555 74  6 M (0.91891892 0.08108108) *
#>        9) V53< 0.00555 20 10 M (0.50000000 0.50000000)  
#>         18) V15< 0.44255 12  2 M (0.83333333 0.16666667) *
#>         19) V15>=0.44255 8  0 R (0.00000000 1.00000000) *
#>      5) V39< 0.10345 9  1 R (0.11111111 0.88888889) *
#>    3) V11< 0.15805 36  3 R (0.08333333 0.91666667) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 R (0.45323741 0.54676259)  
#>   2) V31< 0.4623 57 11 M (0.80701754 0.19298246)  
#>     4) V47>=0.06225 43  2 M (0.95348837 0.04651163) *
#>     5) V47< 0.06225 14  5 R (0.35714286 0.64285714) *
#>   3) V31>=0.4623 82 17 R (0.20731707 0.79268293)  
#>     6) V29>=0.87175 17  5 M (0.70588235 0.29411765) *
#>     7) V29< 0.87175 65  5 R (0.07692308 0.92307692) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V51>=0.01365 79 17 M (0.78481013 0.21518987)  
#>      4) V30>=0.48815 52  2 M (0.96153846 0.03846154) *
#>      5) V30< 0.48815 27 12 R (0.44444444 0.55555556)  
#>       10) V12>=0.39755 10  0 M (1.00000000 0.00000000) *
#>       11) V12< 0.39755 17  2 R (0.11764706 0.88235294) *
#>    3) V51< 0.01365 60 15 R (0.25000000 0.75000000)  
#>      6) V35< 0.1513 12  3 M (0.75000000 0.25000000) *
#>      7) V35>=0.1513 48  6 R (0.12500000 0.87500000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V21>=0.68575 70 13 M (0.81428571 0.18571429)  
#>      4) V43>=0.1096 60  5 M (0.91666667 0.08333333) *
#>      5) V43< 0.1096 10  2 R (0.20000000 0.80000000) *
#>    3) V21< 0.68575 69 16 R (0.23188406 0.76811594)  
#>      6) V12>=0.23195 30 15 M (0.50000000 0.50000000)  
#>       12) V18< 0.3188 14  0 M (1.00000000 0.00000000) *
#>       13) V18>=0.3188 16  1 R (0.06250000 0.93750000) *
#>      7) V12< 0.23195 39  1 R (0.02564103 0.97435897) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V36< 0.4388 78 23 M (0.70512821 0.29487179)  
#>      4) V5>=0.02235 68 14 M (0.79411765 0.20588235)  
#>        8) V46>=0.11 36  0 M (1.00000000 0.00000000) *
#>        9) V46< 0.11 32 14 M (0.56250000 0.43750000)  
#>         18) V36< 0.14875 11  0 M (1.00000000 0.00000000) *
#>         19) V36>=0.14875 21  7 R (0.33333333 0.66666667)  
#>           38) V27< 0.46975 9  2 M (0.77777778 0.22222222) *
#>           39) V27>=0.46975 12  0 R (0.00000000 1.00000000) *
#>      5) V5< 0.02235 10  1 R (0.10000000 0.90000000) *
#>    3) V36>=0.4388 61 13 R (0.21311475 0.78688525)  
#>      6) V31< 0.3668 10  0 M (1.00000000 0.00000000) *
#>      7) V31>=0.3668 51  3 R (0.05882353 0.94117647) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.5755396 0.4244604)  
#>    2) V58>=0.0032 121 41 M (0.6611570 0.3388430)  
#>      4) V4>=0.052 55  6 M (0.8909091 0.1090909) *
#>      5) V4< 0.052 66 31 R (0.4696970 0.5303030)  
#>       10) V23>=0.74665 23  3 M (0.8695652 0.1304348) *
#>       11) V23< 0.74665 43 11 R (0.2558140 0.7441860)  
#>         22) V44< 0.04805 7  0 M (1.0000000 0.0000000) *
#>         23) V44>=0.04805 36  4 R (0.1111111 0.8888889) *
#>    3) V58< 0.0032 18  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V21>=0.5605 81 24 M (0.70370370 0.29629630)  
#>      4) V19< 0.7365 41  2 M (0.95121951 0.04878049) *
#>      5) V19>=0.7365 40 18 R (0.45000000 0.55000000)  
#>       10) V59>=0.0108 15  1 M (0.93333333 0.06666667) *
#>       11) V59< 0.0108 25  4 R (0.16000000 0.84000000)  
#>         22) V2< 0.0329 7  3 M (0.57142857 0.42857143) *
#>         23) V2>=0.0329 18  0 R (0.00000000 1.00000000) *
#>    3) V21< 0.5605 58 13 R (0.22413793 0.77586207)  
#>      6) V44>=0.39015 9  0 M (1.00000000 0.00000000) *
#>      7) V44< 0.39015 49  4 R (0.08163265 0.91836735)  
#>       14) V37< 0.22025 7  3 M (0.57142857 0.42857143) *
#>       15) V37>=0.22025 42  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V52>=0.0076 91 31 M (0.65934066 0.34065934)  
#>      4) V17< 0.6686 68 13 M (0.80882353 0.19117647)  
#>        8) V12>=0.2382 50  2 M (0.96000000 0.04000000) *
#>        9) V12< 0.2382 18  7 R (0.38888889 0.61111111) *
#>      5) V17>=0.6686 23  5 R (0.21739130 0.78260870)  
#>       10) V18>=0.7883 7  2 M (0.71428571 0.28571429) *
#>       11) V18< 0.7883 16  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0076 48  4 R (0.08333333 0.91666667) *
#> 
#> 
#> $weights
#>  [1] 0.6886628 0.4988776 0.7265587 0.5120888 0.6764301 0.8054629 0.9944164
#>  [8] 0.5428706 0.6230120 0.6210972
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0748922 4.6145849
#>   [2,] 1.1875405 5.5019365
#>   [3,] 1.1875405 5.5019365
#>   [4,] 1.1753077 5.5141693
#>   [5,] 3.1967818 3.4926952
#>   [6,] 2.3119681 4.3775089
#>   [7,] 0.5120888 6.1773883
#>   [8,] 2.2206845 4.4687926
#>   [9,] 1.3495707 5.3399063
#>  [10,] 1.9580921 4.7313850
#>  [11,] 2.7380385 3.9514385
#>  [12,] 0.7265587 5.9629184
#>  [13,] 2.6477755 4.0417015
#>  [14,] 2.5351963 4.1542807
#>  [15,] 1.1218896 5.5675874
#>  [16,] 1.8465335 4.8429435
#>  [17,] 1.3495707 5.3399063
#>  [18,] 2.6037826 4.0856944
#>  [19,] 2.0369963 4.6524807
#>  [20,] 3.0780669 3.6114101
#>  [21,] 1.0549593 5.6345177
#>  [22,] 0.6764301 6.0130470
#>  [23,] 0.0000000 6.6894770
#>  [24,] 1.6996292 4.9898478
#>  [25,] 2.4139551 4.2755219
#>  [26,] 2.3534230 4.3360540
#>  [27,] 1.7561979 4.9332791
#>  [28,] 2.6600083 4.0294687
#>  [29,] 0.0000000 6.6894770
#>  [30,] 1.2994421 5.3900350
#>  [31,] 0.0000000 6.6894770
#>  [32,] 1.6708464 5.0186306
#>  [33,] 1.4029887 5.2864883
#>  [34,] 1.6647602 5.0247168
#>  [35,] 1.1218896 5.5675874
#>  [36,] 1.1218896 5.5675874
#>  [37,] 1.6779713 5.0115057
#>  [38,] 1.8237636 4.8657134
#>  [39,] 1.3650929 5.3243841
#>  [40,] 0.6886628 6.0008142
#>  [41,] 0.6230120 6.0664650
#>  [42,] 1.1218896 5.5675874
#>  [43,] 1.1218896 5.5675874
#>  [44,] 1.1658826 5.5235945
#>  [45,] 2.7813961 3.9080809
#>  [46,] 2.2385256 4.4509515
#>  [47,] 1.6174284 5.0720486
#>  [48,] 2.2938584 4.3956186
#>  [49,] 1.0109664 5.6785106
#>  [50,] 1.8164293 4.8730477
#>  [51,] 1.4932940 5.1961830
#>  [52,] 2.9217690 3.7677081
#>  [53,] 2.1768490 4.5126281
#>  [54,] 1.9580921 4.7313850
#>  [55,] 2.6704729 4.0190042
#>  [56,] 1.7181782 4.9712988
#>  [57,] 0.6210972 6.0683798
#>  [58,] 2.7761308 3.9133462
#>  [59,] 2.3226412 4.3668358
#>  [60,] 2.1295172 4.5599599
#>  [61,] 1.7436222 4.9458549
#>  [62,] 1.6830792 5.0063978
#>  [63,] 5.2075840 1.4818930
#>  [64,] 4.6832625 2.0062145
#>  [65,] 4.3687506 2.3207264
#>  [66,] 5.3243841 1.3650929
#>  [67,] 5.0020806 1.6873964
#>  [68,] 4.1665135 2.5229635
#>  [69,] 5.2864883 1.4029887
#>  [70,] 4.3809834 2.3084936
#>  [71,] 4.6531583 2.0363187
#>  [72,] 4.8798610 1.8096160
#>  [73,] 4.7013721 1.9881049
#>  [74,] 5.3919498 1.2975272
#>  [75,] 6.1905994 0.4988776
#>  [76,] 6.0683798 0.6210972
#>  [77,] 5.6345177 1.0549593
#>  [78,] 4.5145429 2.1749341
#>  [79,] 6.0683798 0.6210972
#>  [80,] 5.6950606 0.9944164
#>  [81,] 6.1905994 0.4988776
#>  [82,] 5.3778022 1.3116749
#>  [83,] 6.0008142 0.6886628
#>  [84,] 5.4887254 1.2007516
#>  [85,] 5.0266317 1.6628454
#>  [86,] 6.0683798 0.6210972
#>  [87,] 5.6950606 0.9944164
#>  [88,] 5.5695022 1.1199748
#>  [89,] 4.5618747 2.1276023
#>  [90,] 5.6950606 0.9944164
#>  [91,] 4.8657134 1.8237636
#>  [92,] 6.1773883 0.5120888
#>  [93,] 5.4200478 1.2694292
#>  [94,] 5.5562911 1.1331859
#>  [95,] 6.1466065 0.5428706
#>  [96,] 4.0242695 2.6652075
#>  [97,] 4.5599599 2.1295172
#>  [98,] 6.1773883 0.5120888
#>  [99,] 4.7508282 1.9386489
#> [100,] 6.6894770 0.0000000
#> [101,] 4.5864869 2.1029902
#> [102,] 4.7508282 1.9386489
#> [103,] 4.3975334 2.2919436
#> [104,] 5.5141693 1.1753077
#> [105,] 4.5363582 2.1531188
#> [106,] 5.9629184 0.7265587
#> [107,] 5.0186306 1.6708464
#> [108,] 4.3467271 2.3427499
#> [109,] 5.2075840 1.4818930
#> [110,] 5.5009582 1.1885188
#> [111,] 3.9934877 2.6959893
#> [112,] 4.6647135 2.0247635
#> [113,] 4.0436163 2.6458607
#> [114,] 6.0130470 0.6764301
#> [115,] 4.5065418 2.1829352
#> [116,] 4.6954953 1.9939817
#> [117,] 6.0683798 0.6210972
#> [118,] 5.5255093 1.1639677
#> [119,] 4.8798610 1.8096160
#> [120,] 4.8255065 1.8639705
#> [121,] 6.6894770 0.0000000
#> [122,] 5.6345177 1.0549593
#> [123,] 6.6894770 0.0000000
#> [124,] 4.6533124 2.0361646
#> [125,] 5.6345177 1.0549593
#> [126,] 5.9629184 0.7265587
#> [127,] 6.1466065 0.5428706
#> [128,] 6.6894770 0.0000000
#> [129,] 6.1466065 0.5428706
#> [130,] 6.1466065 0.5428706
#> [131,] 6.1466065 0.5428706
#> [132,] 5.9629184 0.7265587
#> [133,] 6.1466065 0.5428706
#> [134,] 6.1466065 0.5428706
#> [135,] 5.6950606 0.9944164
#> [136,] 6.1466065 0.5428706
#> [137,] 5.4200478 1.2694292
#> [138,] 5.1829719 1.5065051
#> [139,] 4.4564132 2.2330638
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.31017255 0.68982745
#>   [2,] 0.17752367 0.82247633
#>   [3,] 0.17752367 0.82247633
#>   [4,] 0.17569500 0.82430500
#>   [5,] 0.47788217 0.52211783
#>   [6,] 0.34561268 0.65438732
#>   [7,] 0.07655139 0.92344861
#>   [8,] 0.33196683 0.66803317
#>   [9,] 0.20174532 0.79825468
#>  [10,] 0.29271228 0.70728772
#>  [11,] 0.40930532 0.59069468
#>  [12,] 0.10861218 0.89138782
#>  [13,] 0.39581204 0.60418796
#>  [14,] 0.37898274 0.62101726
#>  [15,] 0.16770962 0.83229038
#>  [16,] 0.27603555 0.72396445
#>  [17,] 0.20174532 0.79825468
#>  [18,] 0.38923560 0.61076440
#>  [19,] 0.30450756 0.69549244
#>  [20,] 0.46013565 0.53986435
#>  [21,] 0.15770430 0.84229570
#>  [22,] 0.10111853 0.89888147
#>  [23,] 0.00000000 1.00000000
#>  [24,] 0.25407505 0.74592495
#>  [25,] 0.36085857 0.63914143
#>  [26,] 0.35180972 0.64819028
#>  [27,] 0.26253143 0.73746857
#>  [28,] 0.39764070 0.60235930
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.19425167 0.80574833
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.24977236 0.75022764
#>  [33,] 0.20973070 0.79026930
#>  [34,] 0.24886253 0.75113747
#>  [35,] 0.16770962 0.83229038
#>  [36,] 0.16770962 0.83229038
#>  [37,] 0.25083744 0.74916256
#>  [38,] 0.27263172 0.72736828
#>  [39,] 0.20406571 0.79593429
#>  [40,] 0.10294719 0.89705281
#>  [41,] 0.09313314 0.90686686
#>  [42,] 0.16770962 0.83229038
#>  [43,] 0.16770962 0.83229038
#>  [44,] 0.17428605 0.82571395
#>  [45,] 0.41578678 0.58421322
#>  [46,] 0.33463387 0.66536613
#>  [47,] 0.24178697 0.75821303
#>  [48,] 0.34290550 0.65709450
#>  [49,] 0.15112787 0.84887213
#>  [50,] 0.27153533 0.72846467
#>  [51,] 0.22323031 0.77676969
#>  [52,] 0.43677091 0.56322909
#>  [53,] 0.32541392 0.67458608
#>  [54,] 0.29271228 0.70728772
#>  [55,] 0.39920503 0.60079497
#>  [56,] 0.25684792 0.74315208
#>  [57,] 0.09284690 0.90715310
#>  [58,] 0.41499968 0.58500032
#>  [59,] 0.34720820 0.65279180
#>  [60,] 0.31833836 0.68166164
#>  [61,] 0.26065149 0.73934851
#>  [62,] 0.25160102 0.74839898
#>  [63,] 0.77847401 0.22152599
#>  [64,] 0.70009396 0.29990604
#>  [65,] 0.65307805 0.34692195
#>  [66,] 0.79593429 0.20406571
#>  [67,] 0.74775361 0.25224639
#>  [68,] 0.62284592 0.37715408
#>  [69,] 0.79026930 0.20973070
#>  [70,] 0.65490671 0.34509329
#>  [71,] 0.69559374 0.30440626
#>  [72,] 0.72948319 0.27051681
#>  [73,] 0.70280115 0.29719885
#>  [74,] 0.80603458 0.19396542
#>  [75,] 0.92542352 0.07457648
#>  [76,] 0.90715310 0.09284690
#>  [77,] 0.84229570 0.15770430
#>  [78,] 0.67487232 0.32512768
#>  [79,] 0.90715310 0.09284690
#>  [80,] 0.85134617 0.14865383
#>  [81,] 0.92542352 0.07457648
#>  [82,] 0.80391967 0.19608033
#>  [83,] 0.89705281 0.10294719
#>  [84,] 0.82050142 0.17949858
#>  [85,] 0.75142371 0.24857629
#>  [86,] 0.90715310 0.09284690
#>  [87,] 0.85134617 0.14865383
#>  [88,] 0.83257663 0.16742337
#>  [89,] 0.68194788 0.31805212
#>  [90,] 0.85134617 0.14865383
#>  [91,] 0.72736828 0.27263172
#>  [92,] 0.92344861 0.07655139
#>  [93,] 0.81023491 0.18976509
#>  [94,] 0.83060172 0.16939828
#>  [95,] 0.91884709 0.08115291
#>  [96,] 0.60158208 0.39841792
#>  [97,] 0.68166164 0.31833836
#>  [98,] 0.92344861 0.07655139
#>  [99,] 0.71019426 0.28980574
#> [100,] 1.00000000 0.00000000
#> [101,] 0.68562712 0.31437288
#> [102,] 0.71019426 0.28980574
#> [103,] 0.65738075 0.34261925
#> [104,] 0.82430500 0.17569500
#> [105,] 0.67813347 0.32186653
#> [106,] 0.89138782 0.10861218
#> [107,] 0.75022764 0.24977236
#> [108,] 0.64978579 0.35021421
#> [109,] 0.77847401 0.22152599
#> [110,] 0.82233009 0.17766991
#> [111,] 0.59698055 0.40301945
#> [112,] 0.69732110 0.30267890
#> [113,] 0.60447420 0.39552580
#> [114,] 0.89888147 0.10111853
#> [115,] 0.67367625 0.32632375
#> [116,] 0.70192263 0.29807737
#> [117,] 0.90715310 0.09284690
#> [118,] 0.82600019 0.17399981
#> [119,] 0.72948319 0.27051681
#> [120,] 0.72135781 0.27864219
#> [121,] 1.00000000 0.00000000
#> [122,] 0.84229570 0.15770430
#> [123,] 1.00000000 0.00000000
#> [124,] 0.69561678 0.30438322
#> [125,] 0.84229570 0.15770430
#> [126,] 0.89138782 0.10861218
#> [127,] 0.91884709 0.08115291
#> [128,] 1.00000000 0.00000000
#> [129,] 0.91884709 0.08115291
#> [130,] 0.91884709 0.08115291
#> [131,] 0.91884709 0.08115291
#> [132,] 0.89138782 0.10861218
#> [133,] 0.91884709 0.08115291
#> [134,] 0.91884709 0.08115291
#> [135,] 0.85134617 0.14865383
#> [136,] 0.91884709 0.08115291
#> [137,] 0.81023491 0.18976509
#> [138,] 0.77479478 0.22520522
#> [139,] 0.66618260 0.33381740
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 0.0000000 8.4882517 5.5943718 0.0000000 0.0000000 1.4937531 1.8782261 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 2.3030017 4.2119505 1.9545220 0.6323831 0.0000000 8.8422427 0.0000000 1.8900044 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 2.3312853 1.9081529 0.0000000 1.6837131 0.0000000 1.9825827 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 8.3619400 1.4880487 0.0000000 0.0000000 1.5645287 7.0500938 0.7528370 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 1.9010767 1.7821680 0.0000000 0.0000000 0.0000000 2.1870694 4.0145940 1.2641006 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 2.9383768 1.1862936 0.0000000 0.0000000 2.5761915 0.0000000 4.0692565 3.9940434 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 1.2382088 0.0000000 0.0000000 0.0000000 0.0000000 2.2933903 2.1544208 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 1.1309372 0.0000000 2.8579828 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 77 62 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V21       V11       V31       V36       V12       V18       V51       V44 
#> 8.8422427 8.4882517 8.3619400 7.0500938 5.5943718 4.2119505 4.0692565 4.0145940 
#>       V52       V46        V9        V5       V26       V17       V58       V43 
#> 3.9940434 2.9383768 2.8579828 2.5761915 2.3312853 2.3030017 2.2933903 2.1870694 
#>       V59       V30       V19       V27       V39       V23       V16        V4 
#> 2.1544208 1.9825827 1.9545220 1.9081529 1.9010767 1.8900044 1.8782261 1.7821680 
#>       V29       V35       V15       V32       V45       V53       V47        V7 
#> 1.6837131 1.5645287 1.4937531 1.4880487 1.2641006 1.2382088 1.1862936 1.1309372 
#>       V37        V2        V1       V10       V13       V14       V20       V22 
#> 0.7528370 0.6323831 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V24       V25       V28        V3       V33       V34       V38       V40 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V41       V42       V48       V49       V50       V54       V55       V56 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V57        V6       V60        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203