Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V11>=0.19795 79 16 M (0.79746835 0.20253165)  
#>      4) V3< 0.05315 59  5 M (0.91525424 0.08474576) *
#>      5) V3>=0.05315 20  9 R (0.45000000 0.55000000)  
#>       10) V51>=0.0155 8  0 M (1.00000000 0.00000000) *
#>       11) V51< 0.0155 12  1 R (0.08333333 0.91666667) *
#>    3) V11< 0.19795 60  9 R (0.15000000 0.85000000)  
#>      6) V4>=0.0576 8  3 M (0.62500000 0.37500000) *
#>      7) V4< 0.0576 52  4 R (0.07692308 0.92307692) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.43165468 0.56834532)  
#>    2) V45>=0.2726 31  1 M (0.96774194 0.03225806) *
#>    3) V45< 0.2726 108 30 R (0.27777778 0.72222222)  
#>      6) V36< 0.40335 62 29 R (0.46774194 0.53225806)  
#>       12) V52>=0.0093 29  6 M (0.79310345 0.20689655)  
#>         24) V38>=0.1392 22  1 M (0.95454545 0.04545455) *
#>         25) V38< 0.1392 7  2 R (0.28571429 0.71428571) *
#>       13) V52< 0.0093 33  6 R (0.18181818 0.81818182)  
#>         26) V4>=0.05215 7  2 M (0.71428571 0.28571429) *
#>         27) V4< 0.05215 26  1 R (0.03846154 0.96153846) *
#>      7) V36>=0.40335 46  1 R (0.02173913 0.97826087) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V10>=0.20555 69 17 M (0.75362319 0.24637681)  
#>      4) V51>=0.01285 43  3 M (0.93023256 0.06976744) *
#>      5) V51< 0.01285 26 12 R (0.46153846 0.53846154)  
#>       10) V15< 0.6172 17  5 M (0.70588235 0.29411765) *
#>       11) V15>=0.6172 9  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.20555 70 20 R (0.28571429 0.71428571)  
#>      6) V4>=0.0926 8  0 M (1.00000000 0.00000000) *
#>      7) V4< 0.0926 62 12 R (0.19354839 0.80645161)  
#>       14) V45>=0.1864 16  7 M (0.56250000 0.43750000) *
#>       15) V45< 0.1864 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V36< 0.5047 94 32 M (0.65957447 0.34042553)  
#>      4) V24>=0.4854 70 13 M (0.81428571 0.18571429)  
#>        8) V4>=0.0326 51  3 M (0.94117647 0.05882353) *
#>        9) V4< 0.0326 19  9 R (0.47368421 0.52631579) *
#>      5) V24< 0.4854 24  5 R (0.20833333 0.79166667)  
#>       10) V1>=0.02925 7  2 M (0.71428571 0.28571429) *
#>       11) V1< 0.02925 17  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.5047 45  8 R (0.17777778 0.82222222)  
#>      6) V45>=0.3586 8  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.3586 37  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V28>=0.958 29  1 M (0.96551724 0.03448276) *
#>    3) V28< 0.958 110 44 R (0.40000000 0.60000000)  
#>      6) V23>=0.7553 43 13 M (0.69767442 0.30232558)  
#>       12) V36< 0.50825 32  4 M (0.87500000 0.12500000)  
#>         24) V10>=0.12295 25  0 M (1.00000000 0.00000000) *
#>         25) V10< 0.12295 7  3 R (0.42857143 0.57142857) *
#>       13) V36>=0.50825 11  2 R (0.18181818 0.81818182) *
#>      7) V23< 0.7553 67 14 R (0.20895522 0.79104478)  
#>       14) V44>=0.39625 7  0 M (1.00000000 0.00000000) *
#>       15) V44< 0.39625 60  7 R (0.11666667 0.88333333)  
#>         30) V26< 0.2346 7  2 M (0.71428571 0.28571429) *
#>         31) V26>=0.2346 53  2 R (0.03773585 0.96226415) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 M (0.50359712 0.49640288)  
#>   2) V52>=0.00935 89 27 M (0.69662921 0.30337079)  
#>     4) V13>=0.16265 70 12 M (0.82857143 0.17142857)  
#>       8) V49>=0.01855 62  6 M (0.90322581 0.09677419) *
#>       9) V49< 0.01855 8  2 R (0.25000000 0.75000000) *
#>     5) V13< 0.16265 19  4 R (0.21052632 0.78947368) *
#>   3) V52< 0.00935 50  8 R (0.16000000 0.84000000)  
#>     6) V45>=0.1977 13  6 M (0.53846154 0.46153846) *
#>     7) V45< 0.1977 37  1 R (0.02702703 0.97297297) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V36< 0.4436 89 26 M (0.70786517 0.29213483)  
#>      4) V52>=0.0082 65  9 M (0.86153846 0.13846154)  
#>        8) V58>=0.00335 58  4 M (0.93103448 0.06896552) *
#>        9) V58< 0.00335 7  2 R (0.28571429 0.71428571) *
#>      5) V52< 0.0082 24  7 R (0.29166667 0.70833333)  
#>       10) V36>=0.31835 7  0 M (1.00000000 0.00000000) *
#>       11) V36< 0.31835 17  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.4436 50  9 R (0.18000000 0.82000000)  
#>      6) V45>=0.3586 7  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.3586 43  2 R (0.04651163 0.95348837) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 51 M (0.63309353 0.36690647)  
#>    2) V56< 0.00465 43  2 M (0.95348837 0.04651163) *
#>    3) V56>=0.00465 96 47 R (0.48958333 0.51041667)  
#>      6) V52>=0.0132 32  5 M (0.84375000 0.15625000) *
#>      7) V52< 0.0132 64 20 R (0.31250000 0.68750000)  
#>       14) V57< 0.00415 12  1 M (0.91666667 0.08333333) *
#>       15) V57>=0.00415 52  9 R (0.17307692 0.82692308)  
#>         30) V49>=0.08155 7  1 M (0.85714286 0.14285714) *
#>         31) V49< 0.08155 45  3 R (0.06666667 0.93333333) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V37< 0.47795 106 32 M (0.69811321 0.30188679)  
#>      4) V48>=0.04695 79 11 M (0.86075949 0.13924051)  
#>        8) V9>=0.0724 72  5 M (0.93055556 0.06944444) *
#>        9) V9< 0.0724 7  1 R (0.14285714 0.85714286) *
#>      5) V48< 0.04695 27  6 R (0.22222222 0.77777778)  
#>       10) V48< 0.03795 8  2 M (0.75000000 0.25000000) *
#>       11) V48>=0.03795 19  0 R (0.00000000 1.00000000) *
#>    3) V37>=0.47795 33  4 R (0.12121212 0.87878788)  
#>      6) V37>=0.74275 7  3 M (0.57142857 0.42857143) *
#>      7) V37< 0.74275 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V12>=0.1555 96 30 M (0.68750000 0.31250000)  
#>      4) V23>=0.6987 55  4 M (0.92727273 0.07272727) *
#>      5) V23< 0.6987 41 15 R (0.36585366 0.63414634)  
#>       10) V51>=0.0136 15  3 M (0.80000000 0.20000000) *
#>       11) V51< 0.0136 26  3 R (0.11538462 0.88461538) *
#>    3) V12< 0.1555 43  4 R (0.09302326 0.90697674)  
#>      6) V27< 0.69145 7  3 M (0.57142857 0.42857143) *
#>      7) V27>=0.69145 36  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.8571529 0.7528468 0.7210720 0.9687322 0.6453892 0.8743787
#>  [8] 0.5973669 0.7556690 0.6129926
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.9799608 4.5890788
#>   [2,] 2.3511197 5.2179198
#>   [3,] 2.3288985 5.2401410
#>   [4,] 3.0973710 4.4716686
#>   [5,] 1.5391082 6.0299314
#>   [6,] 2.7466455 4.8223941
#>   [7,] 2.1539051 5.4151345
#>   [8,] 2.9742878 4.5947518
#>   [9,] 2.2274146 5.3416249
#>  [10,] 0.7210720 6.8479675
#>  [11,] 2.1215084 5.4475311
#>  [12,] 1.6272255 5.9418141
#>  [13,] 1.3686616 6.2003779
#>  [14,] 0.5973669 6.9716726
#>  [15,] 2.6085956 4.9604440
#>  [16,] 1.9984252 5.5706143
#>  [17,] 0.6129926 6.9560470
#>  [18,] 2.2073667 5.3616729
#>  [19,] 2.8538325 4.7152070
#>  [20,] 1.5362860 6.0327536
#>  [21,] 2.3934389 5.1756007
#>  [22,] 1.2103595 6.3586800
#>  [23,] 1.3184390 6.2506006
#>  [24,] 1.8557488 5.7132908
#>  [25,] 2.6391879 4.9298517
#>  [26,] 1.5045112 6.0645284
#>  [27,] 1.2103595 6.3586800
#>  [28,] 0.0000000 7.5690396
#>  [29,] 0.7556690 6.8133705
#>  [30,] 2.0712858 5.4977538
#>  [31,] 0.0000000 7.5690396
#>  [32,] 2.7494677 4.8195718
#>  [33,] 2.1336529 5.4353867
#>  [34,] 2.8683816 4.7006580
#>  [35,] 0.6129926 6.9560470
#>  [36,] 1.3530360 6.2160036
#>  [37,] 0.5973669 6.9716726
#>  [38,] 1.3530360 6.2160036
#>  [39,] 0.7556690 6.8133705
#>  [40,] 2.3217681 5.2472714
#>  [41,] 1.3530360 6.2160036
#>  [42,] 0.0000000 7.5690396
#>  [43,] 0.5973669 6.9716726
#>  [44,] 2.3217681 5.2472714
#>  [45,] 0.5973669 6.9716726
#>  [46,] 1.5660991 6.0029405
#>  [47,] 1.9660286 5.6030110
#>  [48,] 2.3373938 5.2316458
#>  [49,] 1.7244012 5.8446384
#>  [50,] 0.9687322 6.6003074
#>  [51,] 0.5973669 6.9716726
#>  [52,] 1.3530360 6.2160036
#>  [53,] 2.1928176 5.3762219
#>  [54,] 1.9660286 5.6030110
#>  [55,] 1.4767411 6.0922985
#>  [56,] 2.7002638 4.8687758
#>  [57,] 1.6898042 5.8792354
#>  [58,] 2.2754369 5.2936026
#>  [59,] 1.9314316 5.6376080
#>  [60,] 3.5185621 4.0504775
#>  [61,] 1.4010583 6.1679813
#>  [62,] 2.2601802 5.3088593
#>  [63,] 2.7790421 4.7899974
#>  [64,] 1.9956030 5.5734366
#>  [65,] 2.7494677 4.8195718
#>  [66,] 2.1364751 5.4325644
#>  [67,] 1.5045112 6.0645284
#>  [68,] 5.7259288 1.8431108
#>  [69,] 5.3575512 2.2114883
#>  [70,] 6.6003074 0.9687322
#>  [71,] 7.5690396 0.0000000
#>  [72,] 5.3136506 2.2553890
#>  [73,] 6.1726078 1.3964317
#>  [74,] 5.9549182 1.6141214
#>  [75,] 5.2472714 2.3217681
#>  [76,] 4.8687758 2.7002638
#>  [77,] 5.2401410 2.3288985
#>  [78,] 6.8161928 0.7528468
#>  [79,] 6.9716726 0.5973669
#>  [80,] 6.2506006 1.3184390
#>  [81,] 6.8479675 0.7210720
#>  [82,] 6.1402112 1.4288284
#>  [83,] 7.5690396 0.0000000
#>  [84,] 6.8479675 0.7210720
#>  [85,] 6.0492717 1.5197679
#>  [86,] 7.5690396 0.0000000
#>  [87,] 7.5690396 0.0000000
#>  [88,] 5.8446384 1.7244012
#>  [89,] 5.8446384 1.7244012
#>  [90,] 6.6003074 0.9687322
#>  [91,] 6.1402112 1.4288284
#>  [92,] 6.7118866 0.8571529
#>  [93,] 4.9874854 2.5815541
#>  [94,] 6.0645284 1.5045112
#>  [95,] 6.6003074 0.9687322
#>  [96,] 7.5690396 0.0000000
#>  [97,] 7.5690396 0.0000000
#>  [98,] 7.5690396 0.0000000
#>  [99,] 5.8446384 1.7244012
#> [100,] 6.6946609 0.8743787
#> [101,] 5.3873644 2.1816752
#> [102,] 6.2506006 1.3184390
#> [103,] 6.2025783 1.3664613
#> [104,] 5.9284475 1.6405921
#> [105,] 5.3259993 2.2430403
#> [106,] 5.2830582 2.2859813
#> [107,] 6.2003779 1.3686616
#> [108,] 6.8133705 0.7556690
#> [109,] 6.9560470 0.6129926
#> [110,] 7.5690396 0.0000000
#> [111,] 6.8133705 0.7556690
#> [112,] 6.6946609 0.8743787
#> [113,] 7.5690396 0.0000000
#> [114,] 5.8168683 1.7521713
#> [115,] 6.1726078 1.3964317
#> [116,] 5.9735889 1.5954507
#> [117,] 6.6003074 0.9687322
#> [118,] 5.2662428 2.3027968
#> [119,] 5.9735889 1.5954507
#> [120,] 6.2160036 1.3530360
#> [121,] 6.0664974 1.5025422
#> [122,] 4.6986890 2.8703506
#> [123,] 5.7431545 1.8258851
#> [124,] 5.9735889 1.5954507
#> [125,] 7.5690396 0.0000000
#> [126,] 6.8479675 0.7210720
#> [127,] 5.3605963 2.2084433
#> [128,] 6.6003074 0.9687322
#> [129,] 5.9908146 1.5782250
#> [130,] 6.7118866 0.8571529
#> [131,] 6.0951207 1.4739189
#> [132,] 5.3454254 2.2236142
#> [133,] 5.8375080 1.7315316
#> [134,] 5.8375080 1.7315316
#> [135,] 5.8375080 1.7315316
#> [136,] 6.8479675 0.7210720
#> [137,] 6.8479675 0.7210720
#> [138,] 5.9549182 1.6141214
#> [139,] 6.2025783 1.3664613
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.39370395 0.60629605
#>   [2,] 0.31062326 0.68937674
#>   [3,] 0.30768746 0.69231254
#>   [4,] 0.40921586 0.59078414
#>   [5,] 0.20334260 0.79665740
#>   [6,] 0.36287900 0.63712100
#>   [7,] 0.28456782 0.71543218
#>   [8,] 0.39295445 0.60704555
#>   [9,] 0.29427969 0.70572031
#>  [10,] 0.09526599 0.90473401
#>  [11,] 0.28028767 0.71971233
#>  [12,] 0.21498441 0.78501559
#>  [13,] 0.18082369 0.81917631
#>  [14,] 0.07892242 0.92107758
#>  [15,] 0.34464023 0.65535977
#>  [16,] 0.26402626 0.73597374
#>  [17,] 0.08098684 0.91901316
#>  [18,] 0.29163102 0.70836898
#>  [19,] 0.37704024 0.62295976
#>  [20,] 0.20296974 0.79703026
#>  [21,] 0.31621435 0.68378565
#>  [22,] 0.15990926 0.84009074
#>  [23,] 0.17418841 0.82581159
#>  [24,] 0.24517625 0.75482375
#>  [25,] 0.34868201 0.65131799
#>  [26,] 0.19877174 0.80122826
#>  [27,] 0.15990926 0.84009074
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.09983685 0.90016315
#>  [30,] 0.27365239 0.72634761
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.36325186 0.63674814
#>  [33,] 0.28189216 0.71810784
#>  [34,] 0.37896242 0.62103758
#>  [35,] 0.08098684 0.91901316
#>  [36,] 0.17875927 0.82124073
#>  [37,] 0.07892242 0.92107758
#>  [38,] 0.17875927 0.82124073
#>  [39,] 0.09983685 0.90016315
#>  [40,] 0.30674541 0.69325459
#>  [41,] 0.17875927 0.82124073
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.07892242 0.92107758
#>  [44,] 0.30674541 0.69325459
#>  [45,] 0.07892242 0.92107758
#>  [46,] 0.20690856 0.79309144
#>  [47,] 0.25974611 0.74025389
#>  [48,] 0.30880983 0.69119017
#>  [49,] 0.22782299 0.77217701
#>  [50,] 0.12798614 0.87201386
#>  [51,] 0.07892242 0.92107758
#>  [52,] 0.17875927 0.82124073
#>  [53,] 0.28970884 0.71029116
#>  [54,] 0.25974611 0.74025389
#>  [55,] 0.19510284 0.80489716
#>  [56,] 0.35675117 0.64324883
#>  [57,] 0.22325213 0.77674787
#>  [58,] 0.30062426 0.69937574
#>  [59,] 0.25517525 0.74482475
#>  [60,] 0.46486242 0.53513758
#>  [61,] 0.18510384 0.81489616
#>  [62,] 0.29860859 0.70139141
#>  [63,] 0.36715915 0.63284085
#>  [64,] 0.26365340 0.73634660
#>  [65,] 0.36325186 0.63674814
#>  [66,] 0.28226502 0.71773498
#>  [67,] 0.19877174 0.80122826
#>  [68,] 0.75649344 0.24350656
#>  [69,] 0.70782445 0.29217555
#>  [70,] 0.87201386 0.12798614
#>  [71,] 1.00000000 0.00000000
#>  [72,] 0.70202442 0.29797558
#>  [73,] 0.81550741 0.18449259
#>  [74,] 0.78674687 0.21325313
#>  [75,] 0.69325459 0.30674541
#>  [76,] 0.64324883 0.35675117
#>  [77,] 0.69231254 0.30768746
#>  [78,] 0.90053602 0.09946398
#>  [79,] 0.92107758 0.07892242
#>  [80,] 0.82581159 0.17418841
#>  [81,] 0.90473401 0.09526599
#>  [82,] 0.81122725 0.18877275
#>  [83,] 1.00000000 0.00000000
#>  [84,] 0.90473401 0.09526599
#>  [85,] 0.79921258 0.20078742
#>  [86,] 1.00000000 0.00000000
#>  [87,] 1.00000000 0.00000000
#>  [88,] 0.77217701 0.22782299
#>  [89,] 0.77217701 0.22782299
#>  [90,] 0.87201386 0.12798614
#>  [91,] 0.81122725 0.18877275
#>  [92,] 0.88675539 0.11324461
#>  [93,] 0.65893240 0.34106760
#>  [94,] 0.80122826 0.19877174
#>  [95,] 0.87201386 0.12798614
#>  [96,] 1.00000000 0.00000000
#>  [97,] 1.00000000 0.00000000
#>  [98,] 1.00000000 0.00000000
#>  [99,] 0.77217701 0.22782299
#> [100,] 0.88447958 0.11552042
#> [101,] 0.71176327 0.28823673
#> [102,] 0.82581159 0.17418841
#> [103,] 0.81946702 0.18053298
#> [104,] 0.78324963 0.21675037
#> [105,] 0.70365589 0.29634411
#> [106,] 0.69798264 0.30201736
#> [107,] 0.81917631 0.18082369
#> [108,] 0.90016315 0.09983685
#> [109,] 0.91901316 0.08098684
#> [110,] 1.00000000 0.00000000
#> [111,] 0.90016315 0.09983685
#> [112,] 0.88447958 0.11552042
#> [113,] 1.00000000 0.00000000
#> [114,] 0.76850811 0.23149189
#> [115,] 0.81550741 0.18449259
#> [116,] 0.78921359 0.21078641
#> [117,] 0.87201386 0.12798614
#> [118,] 0.69576103 0.30423897
#> [119,] 0.78921359 0.21078641
#> [120,] 0.82124073 0.17875927
#> [121,] 0.80148840 0.19851160
#> [122,] 0.62077744 0.37922256
#> [123,] 0.75876925 0.24123075
#> [124,] 0.78921359 0.21078641
#> [125,] 1.00000000 0.00000000
#> [126,] 0.90473401 0.09526599
#> [127,] 0.70822675 0.29177325
#> [128,] 0.87201386 0.12798614
#> [129,] 0.79148940 0.20851060
#> [130,] 0.88675539 0.11324461
#> [131,] 0.80527003 0.19472997
#> [132,] 0.70622241 0.29377759
#> [133,] 0.77123496 0.22876504
#> [134,] 0.77123496 0.22876504
#> [135,] 0.77123496 0.22876504
#> [136,] 0.90473401 0.09526599
#> [137,] 0.90473401 0.09526599
#> [138,] 0.78674687 0.21325313
#> [139,] 0.81946702 0.18053298
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.9729634  3.9776055  5.9737098  3.4315321  1.9649971  0.0000000  1.1774096 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  5.6530090  2.5239183  0.0000000  1.4623934  0.6256759  3.7923596 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  1.3510738  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000 13.6236635  4.1016639  1.0860561  0.0000000  5.0360018  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  2.5273516 12.3696206  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  4.5833049  2.2468540  0.0000000  0.0000000  4.5723931 10.3837329  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  2.0363971  1.7177138  1.2131122  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  1.5954869 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V36        V45        V52        V11        V23         V4        V48 
#> 13.6236635 12.3696206 10.3837329  5.9737098  5.6530090  5.0360018  4.5833049 
#>        V51        V37        V10        V28        V12        V44        V24 
#>  4.5723931  4.1016639  3.9776055  3.7923596  3.4315321  2.5273516  2.5239183 
#>        V49        V56        V13        V57         V9        V26         V3 
#>  2.2468540  2.0363971  1.9649971  1.7177138  1.5954869  1.4623934  1.3510738 
#>        V58        V15        V38         V1        V27        V14        V16 
#>  1.2131122  1.1774096  1.0860561  0.9729634  0.6256759  0.0000000  0.0000000 
#>        V17        V18        V19         V2        V20        V21        V22 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V25        V29        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V39        V40        V41        V42        V43        V46 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V47         V5        V50        V53        V54        V55        V59 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V6        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913