Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 M (0.52517986 0.47482014)  
#>   2) V11>=0.19795 86 21 M (0.75581395 0.24418605)  
#>     4) V39>=0.161 71 10 M (0.85915493 0.14084507)  
#>       8) V6< 0.1809 60  3 M (0.95000000 0.05000000) *
#>       9) V6>=0.1809 11  4 R (0.36363636 0.63636364) *
#>     5) V39< 0.161 15  4 R (0.26666667 0.73333333) *
#>   3) V11< 0.19795 53  8 R (0.15094340 0.84905660)  
#>     6) V45>=0.16055 11  5 M (0.54545455 0.45454545) *
#>     7) V45< 0.16055 42  2 R (0.04761905 0.95238095) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.18875 85 20 M (0.76470588 0.23529412)  
#>      4) V53>=0.00475 72 10 M (0.86111111 0.13888889)  
#>        8) V37< 0.3501 50  1 M (0.98000000 0.02000000) *
#>        9) V37>=0.3501 22  9 M (0.59090909 0.40909091)  
#>         18) V17< 0.46615 11  0 M (1.00000000 0.00000000) *
#>         19) V17>=0.46615 11  2 R (0.18181818 0.81818182) *
#>      5) V53< 0.00475 13  3 R (0.23076923 0.76923077) *
#>    3) V11< 0.18875 54  9 R (0.16666667 0.83333333)  
#>      6) V4>=0.05095 9  2 M (0.77777778 0.22222222) *
#>      7) V4< 0.05095 45  2 R (0.04444444 0.95555556) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V10>=0.1279 98 29 M (0.70408163 0.29591837)  
#>      4) V49>=0.01745 86 17 M (0.80232558 0.19767442)  
#>        8) V52>=0.00895 69  6 M (0.91304348 0.08695652)  
#>         16) V3< 0.07895 62  2 M (0.96774194 0.03225806) *
#>         17) V3>=0.07895 7  3 R (0.42857143 0.57142857) *
#>        9) V52< 0.00895 17  6 R (0.35294118 0.64705882) *
#>      5) V49< 0.01745 12  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.1279 41  3 R (0.07317073 0.92682927) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 M (0.60431655 0.39568345)  
#>    2) V51>=0.0136 72 11 M (0.84722222 0.15277778)  
#>      4) V36< 0.52685 55  2 M (0.96363636 0.03636364) *
#>      5) V36>=0.52685 17  8 R (0.47058824 0.52941176) *
#>    3) V51< 0.0136 67 23 R (0.34328358 0.65671642)  
#>      6) V56>=0.0071 32 12 M (0.62500000 0.37500000)  
#>       12) V52>=0.00775 21  2 M (0.90476190 0.09523810) *
#>       13) V52< 0.00775 11  1 R (0.09090909 0.90909091) *
#>      7) V56< 0.0071 35  3 R (0.08571429 0.91428571) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 M (0.61870504 0.38129496)  
#>    2) V11>=0.1707 99 20 M (0.79797980 0.20202020)  
#>      4) V12>=0.12505 91 13 M (0.85714286 0.14285714)  
#>        8) V34< 0.6644 84  8 M (0.90476190 0.09523810)  
#>         16) V7< 0.21065 77  3 M (0.96103896 0.03896104) *
#>         17) V7>=0.21065 7  2 R (0.28571429 0.71428571) *
#>        9) V34>=0.6644 7  2 R (0.28571429 0.71428571) *
#>      5) V12< 0.12505 8  1 R (0.12500000 0.87500000) *
#>    3) V11< 0.1707 40  7 R (0.17500000 0.82500000)  
#>      6) V4>=0.05205 7  0 M (1.00000000 0.00000000) *
#>      7) V4< 0.05205 33  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V49>=0.0509 60 10 M (0.83333333 0.16666667)  
#>      4) V39>=0.17185 51  4 M (0.92156863 0.07843137)  
#>        8) V5>=0.0439 44  0 M (1.00000000 0.00000000) *
#>        9) V5< 0.0439 7  3 R (0.42857143 0.57142857) *
#>      5) V39< 0.17185 9  3 R (0.33333333 0.66666667) *
#>    3) V49< 0.0509 79 29 R (0.36708861 0.63291139)  
#>      6) V26< 0.5753 13  0 M (1.00000000 0.00000000) *
#>      7) V26>=0.5753 66 16 R (0.24242424 0.75757576)  
#>       14) V50< 0.00665 7  0 M (1.00000000 0.00000000) *
#>       15) V50>=0.00665 59  9 R (0.15254237 0.84745763)  
#>         30) V15< 0.11695 8  2 M (0.75000000 0.25000000) *
#>         31) V15>=0.11695 51  3 R (0.05882353 0.94117647) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V9>=0.11575 92 36 M (0.60869565 0.39130435)  
#>      4) V17< 0.49925 52  8 M (0.84615385 0.15384615)  
#>        8) V14>=0.1304 44  2 M (0.95454545 0.04545455) *
#>        9) V14< 0.1304 8  2 R (0.25000000 0.75000000) *
#>      5) V17>=0.49925 40 12 R (0.30000000 0.70000000)  
#>       10) V31< 0.3674 13  3 M (0.76923077 0.23076923) *
#>       11) V31>=0.3674 27  2 R (0.07407407 0.92592593) *
#>    3) V9< 0.11575 47  7 R (0.14893617 0.85106383)  
#>      6) V24>=0.89695 7  2 M (0.71428571 0.28571429) *
#>      7) V24< 0.89695 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V51>=0.01285 71 13 M (0.81690141 0.18309859)  
#>      4) V11>=0.1957 53  3 M (0.94339623 0.05660377) *
#>      5) V11< 0.1957 18  8 R (0.44444444 0.55555556) *
#>    3) V51< 0.01285 68 21 R (0.30882353 0.69117647)  
#>      6) V40>=0.3933 17  4 M (0.76470588 0.23529412) *
#>      7) V40< 0.3933 51  8 R (0.15686275 0.84313725)  
#>       14) V26>=0.9689 7  2 M (0.71428571 0.28571429) *
#>       15) V26< 0.9689 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V2>=0.03765 61 14 M (0.77049180 0.22950820)  
#>      4) V51>=0.0088 53  7 M (0.86792453 0.13207547) *
#>      5) V51< 0.0088 8  1 R (0.12500000 0.87500000) *
#>    3) V2< 0.03765 78 25 R (0.32051282 0.67948718)  
#>      6) V23>=0.79085 27  9 M (0.66666667 0.33333333)  
#>       12) V39>=0.19685 18  1 M (0.94444444 0.05555556) *
#>       13) V39< 0.19685 9  1 R (0.11111111 0.88888889) *
#>      7) V23< 0.79085 51  7 R (0.13725490 0.86274510)  
#>       14) V46>=0.22415 8  2 M (0.75000000 0.25000000) *
#>       15) V46< 0.22415 43  1 R (0.02325581 0.97674419) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V12>=0.1677 104 34 M (0.67307692 0.32692308)  
#>      4) V5>=0.02145 90 20 M (0.77777778 0.22222222)  
#>        8) V23>=0.633 58  4 M (0.93103448 0.06896552) *
#>        9) V23< 0.633 32 16 M (0.50000000 0.50000000)  
#>         18) V49>=0.0615 12  0 M (1.00000000 0.00000000) *
#>         19) V49< 0.0615 20  4 R (0.20000000 0.80000000) *
#>      5) V5< 0.02145 14  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.1677 35  5 R (0.14285714 0.85714286)  
#>      6) V20>=0.79015 7  2 M (0.71428571 0.28571429) *
#>      7) V20< 0.79015 28  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.6665923 0.9180001 0.6306167 0.6016652 0.8178038 0.5501568 0.7208999
#>  [8] 0.6327651 0.7139325 0.9588317
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.2167491 5.9945150
#>   [2,] 1.8657546 5.3455096
#>   [3,] 1.7845871 5.4266770
#>   [4,] 1.9821900 5.2290742
#>   [5,] 2.1983286 5.0129356
#>   [6,] 2.2335978 4.9776663
#>   [7,] 1.5894485 5.6218157
#>   [8,] 1.2710567 5.9402074
#>   [9,] 0.5501568 6.6611073
#>  [10,] 0.6016652 6.6095990
#>  [11,] 1.1829219 6.0283422
#>  [12,] 1.8968545 5.3144097
#>  [13,] 3.5447787 3.6664855
#>  [14,] 2.5634467 4.6478174
#>  [15,] 0.6665923 6.5446719
#>  [16,] 1.2640894 5.9471748
#>  [17,] 1.2167491 5.9945150
#>  [18,] 2.9998932 4.2113709
#>  [19,] 2.2813968 4.9298673
#>  [20,] 1.4843960 5.7268681
#>  [21,] 2.4567038 4.7545604
#>  [22,] 0.9180001 6.2932641
#>  [23,] 1.5604969 5.6507672
#>  [24,] 1.3155977 5.8956664
#>  [25,] 0.7208999 6.4903643
#>  [26,] 0.6016652 6.6095990
#>  [27,] 1.3225651 5.8886991
#>  [28,] 0.0000000 7.2112642
#>  [29,] 1.2993573 5.9119068
#>  [30,] 1.9299741 5.2812901
#>  [31,] 1.9306816 5.2805825
#>  [32,] 2.5095969 4.7016672
#>  [33,] 2.5634467 4.6478174
#>  [34,] 2.1106538 5.1006104
#>  [35,] 0.6327651 6.5784991
#>  [36,] 0.0000000 7.2112642
#>  [37,] 0.0000000 7.2112642
#>  [38,] 1.5387037 5.6725605
#>  [39,] 1.4505688 5.7606953
#>  [40,] 0.5501568 6.6611073
#>  [41,] 1.1829219 6.0283422
#>  [42,] 0.6327651 6.5784991
#>  [43,] 0.7139325 6.4973316
#>  [44,] 1.2640894 5.9471748
#>  [45,] 0.5501568 6.6611073
#>  [46,] 0.5501568 6.6611073
#>  [47,] 1.1829219 6.0283422
#>  [48,] 1.9306816 5.2805825
#>  [49,] 0.6327651 6.5784991
#>  [50,] 0.0000000 7.2112642
#>  [51,] 0.6016652 6.6095990
#>  [52,] 0.9588317 6.2524324
#>  [53,] 3.2963008 3.9149634
#>  [54,] 2.2813968 4.9298673
#>  [55,] 2.1755809 5.0356833
#>  [56,] 3.3252523 3.8860118
#>  [57,] 1.5317363 5.6795279
#>  [58,] 1.7358039 5.4754603
#>  [59,] 1.4843960 5.7268681
#>  [60,] 1.3466976 5.8645665
#>  [61,] 1.3445493 5.8667149
#>  [62,] 1.9842817 5.2269824
#>  [63,] 1.2993573 5.9119068
#>  [64,] 0.0000000 7.2112642
#>  [65,] 2.2405652 4.9706990
#>  [66,] 1.3515166 5.8597475
#>  [67,] 2.7341898 4.4770744
#>  [68,] 1.9531818 5.2580823
#>  [69,] 1.1829219 6.0283422
#>  [70,] 5.3462172 1.8650470
#>  [71,] 5.2339498 1.9773143
#>  [72,] 5.6218157 1.5894485
#>  [73,] 5.5384999 1.6727643
#>  [74,] 6.5806474 0.6306167
#>  [75,] 5.1458150 2.0654492
#>  [76,] 3.9877347 3.2235295
#>  [77,] 5.2269824 1.9842817
#>  [78,] 4.5603902 2.6508740
#>  [79,] 5.3165581 1.8947061
#>  [80,] 5.2812901 1.9299741
#>  [81,] 6.0304906 1.1807736
#>  [82,] 5.6915989 1.5196653
#>  [83,] 5.7431072 1.4681569
#>  [84,] 5.8237720 1.3874922
#>  [85,] 4.9417474 2.2695167
#>  [86,] 7.2112642 0.0000000
#>  [87,] 5.2736151 1.9376490
#>  [88,] 5.2262749 1.9849893
#>  [89,] 4.3089823 2.9022818
#>  [90,] 5.8575992 1.3536650
#>  [91,] 6.4903643 0.7208999
#>  [92,] 5.2126868 1.9985773
#>  [93,] 6.0304906 1.1807736
#>  [94,] 5.9402074 1.2710567
#>  [95,] 5.7628437 1.4484205
#>  [96,] 6.6611073 0.5501568
#>  [97,] 6.5806474 0.6306167
#>  [98,] 6.5446719 0.6665923
#>  [99,] 5.9789822 1.2322819
#> [100,] 5.7917952 1.4194689
#> [101,] 4.8329635 2.3783007
#> [102,] 4.6110216 2.6002425
#> [103,] 5.0201505 2.1911137
#> [104,] 5.0716588 2.1396053
#> [105,] 4.8426952 2.3685689
#> [106,] 5.7268681 1.4843960
#> [107,] 7.2112642 0.0000000
#> [108,] 6.4903643 0.7208999
#> [109,] 5.6218157 1.5894485
#> [110,] 6.2932641 0.9180001
#> [111,] 4.7057771 2.5054870
#> [112,] 4.7127445 2.4985196
#> [113,] 5.0708953 2.1403689
#> [114,] 6.6095990 0.6016652
#> [115,] 7.2112642 0.0000000
#> [116,] 6.6095990 0.6016652
#> [117,] 6.4973316 0.7139325
#> [118,] 5.7628437 1.4484205
#> [119,] 5.6196673 1.5915968
#> [120,] 6.6095990 0.6016652
#> [121,] 4.6253173 2.5859469
#> [122,] 5.9789822 1.2322819
#> [123,] 5.9789822 1.2322819
#> [124,] 5.4288254 1.7824388
#> [125,] 5.6266718 1.5845924
#> [126,] 5.5793315 1.6319326
#> [127,] 7.2112642 0.0000000
#> [128,] 5.7268681 1.4843960
#> [129,] 5.9789822 1.2322819
#> [130,] 5.2805825 1.9306816
#> [131,] 4.8176000 2.3936642
#> [132,] 5.2650497 1.9462144
#> [133,] 5.9945150 1.2167491
#> [134,] 5.2805825 1.9306816
#> [135,] 5.0765149 2.1347492
#> [136,] 5.9945150 1.2167491
#> [137,] 6.2932641 0.9180001
#> [138,] 5.2650497 1.9462144
#> [139,] 5.0222073 2.1890569
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.16872896 0.83127104
#>   [2,] 0.25872781 0.74127219
#>   [3,] 0.24747216 0.75252784
#>   [4,] 0.27487413 0.72512587
#>   [5,] 0.30484649 0.69515351
#>   [6,] 0.30973735 0.69026265
#>   [7,] 0.22041191 0.77958809
#>   [8,] 0.17625991 0.82374009
#>   [9,] 0.07629132 0.92370868
#>  [10,] 0.08343408 0.91656592
#>  [11,] 0.16403808 0.83596192
#>  [12,] 0.26304049 0.73695951
#>  [13,] 0.49156134 0.50843866
#>  [14,] 0.35547813 0.64452187
#>  [15,] 0.09243764 0.90756236
#>  [16,] 0.17529373 0.82470627
#>  [17,] 0.16872896 0.83127104
#>  [18,] 0.41600102 0.58399898
#>  [19,] 0.31636573 0.68363427
#>  [20,] 0.20584408 0.79415592
#>  [21,] 0.34067588 0.65932412
#>  [22,] 0.12730086 0.87269914
#>  [23,] 0.21639714 0.78360286
#>  [24,] 0.18243649 0.81756351
#>  [25,] 0.09996859 0.90003141
#>  [26,] 0.08343408 0.91656592
#>  [27,] 0.18340267 0.81659733
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.18018441 0.81981559
#>  [30,] 0.26763325 0.73236675
#>  [31,] 0.26773137 0.73226863
#>  [32,] 0.34801068 0.65198932
#>  [33,] 0.35547813 0.64452187
#>  [34,] 0.29268846 0.70731154
#>  [35,] 0.08774676 0.91225324
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.21337502 0.78662498
#>  [39,] 0.20115320 0.79884680
#>  [40,] 0.07629132 0.92370868
#>  [41,] 0.16403808 0.83596192
#>  [42,] 0.08774676 0.91225324
#>  [43,] 0.09900241 0.90099759
#>  [44,] 0.17529373 0.82470627
#>  [45,] 0.07629132 0.92370868
#>  [46,] 0.07629132 0.92370868
#>  [47,] 0.16403808 0.83596192
#>  [48,] 0.26773137 0.73226863
#>  [49,] 0.08774676 0.91225324
#>  [50,] 0.00000000 1.00000000
#>  [51,] 0.08343408 0.91656592
#>  [52,] 0.13296306 0.86703694
#>  [53,] 0.45710443 0.54289557
#>  [54,] 0.31636573 0.68363427
#>  [55,] 0.30169202 0.69830798
#>  [56,] 0.46111920 0.53888080
#>  [57,] 0.21240884 0.78759116
#>  [58,] 0.24070729 0.75929271
#>  [59,] 0.20584408 0.79415592
#>  [60,] 0.18674917 0.81325083
#>  [61,] 0.18645126 0.81354874
#>  [62,] 0.27516420 0.72483580
#>  [63,] 0.18018441 0.81981559
#>  [64,] 0.00000000 1.00000000
#>  [65,] 0.31070353 0.68929647
#>  [66,] 0.18741744 0.81258256
#>  [67,] 0.37915541 0.62084459
#>  [68,] 0.27085152 0.72914848
#>  [69,] 0.16403808 0.83596192
#>  [70,] 0.74137031 0.25862969
#>  [71,] 0.72580198 0.27419802
#>  [72,] 0.77958809 0.22041191
#>  [73,] 0.76803453 0.23196547
#>  [74,] 0.91255115 0.08744885
#>  [75,] 0.71358015 0.28641985
#>  [76,] 0.55298691 0.44701309
#>  [77,] 0.72483580 0.27516420
#>  [78,] 0.63239816 0.36760184
#>  [79,] 0.73725743 0.26274257
#>  [80,] 0.73236675 0.26763325
#>  [81,] 0.83625984 0.16374016
#>  [82,] 0.78926506 0.21073494
#>  [83,] 0.79640783 0.20359217
#>  [84,] 0.80759377 0.19240623
#>  [85,] 0.68528171 0.31471829
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.73130245 0.26869755
#>  [88,] 0.72473768 0.27526232
#>  [89,] 0.59753495 0.40246505
#>  [90,] 0.81228465 0.18771535
#>  [91,] 0.90003141 0.09996859
#>  [92,] 0.72285340 0.27714660
#>  [93,] 0.83625984 0.16374016
#>  [94,] 0.82374009 0.17625991
#>  [95,] 0.79914472 0.20085528
#>  [96,] 0.92370868 0.07629132
#>  [97,] 0.91255115 0.08744885
#>  [98,] 0.90756236 0.09243764
#>  [99,] 0.82911707 0.17088293
#> [100,] 0.80315949 0.19684051
#> [101,] 0.67019643 0.32980357
#> [102,] 0.63941932 0.36058068
#> [103,] 0.69615401 0.30384599
#> [104,] 0.70329678 0.29670322
#> [105,] 0.67154595 0.32845405
#> [106,] 0.79415592 0.20584408
#> [107,] 1.00000000 0.00000000
#> [108,] 0.90003141 0.09996859
#> [109,] 0.77958809 0.22041191
#> [110,] 0.87269914 0.12730086
#> [111,] 0.65255925 0.34744075
#> [112,] 0.65352543 0.34647457
#> [113,] 0.70319090 0.29680910
#> [114,] 0.91656592 0.08343408
#> [115,] 1.00000000 0.00000000
#> [116,] 0.91656592 0.08343408
#> [117,] 0.90099759 0.09900241
#> [118,] 0.79914472 0.20085528
#> [119,] 0.77929018 0.22070982
#> [120,] 0.91656592 0.08343408
#> [121,] 0.64140172 0.35859828
#> [122,] 0.82911707 0.17088293
#> [123,] 0.82911707 0.17088293
#> [124,] 0.75282576 0.24717424
#> [125,] 0.78026150 0.21973850
#> [126,] 0.77369673 0.22630327
#> [127,] 1.00000000 0.00000000
#> [128,] 0.79415592 0.20584408
#> [129,] 0.82911707 0.17088293
#> [130,] 0.73226863 0.26773137
#> [131,] 0.66806594 0.33193406
#> [132,] 0.73011466 0.26988534
#> [133,] 0.83127104 0.16872896
#> [134,] 0.73226863 0.26773137
#> [135,] 0.70397018 0.29602982
#> [136,] 0.83127104 0.16872896
#> [137,] 0.87269914 0.12730086
#> [138,] 0.73011466 0.26988534
#> [139,] 0.69643924 0.30356076
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  4.1738506 17.2557032  5.9148024  0.0000000  1.3933895  1.0454729 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  4.7407573  0.0000000  0.0000000  2.8463709  1.5758489  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  4.1452083  1.0901473  0.0000000  2.8902310  0.0000000  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.6632826  0.0000000  1.7584407  0.0000000  0.0000000  1.1648590 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  1.0925790  1.2213645  0.0000000  4.2157521  4.8465381  1.7146518 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.8283966  1.4630150  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  0.0000000  7.4523847  4.6663889  1.4222267  7.8894098  3.2069852  2.3104267 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  1.6825817  0.0000000  0.0000000  0.0000000  1.2255193 
#>        V60         V7         V8         V9 
#>  0.0000000  1.3766516  0.0000000  2.7267635 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 70 69 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V51        V49        V12         V4        V17         V5 
#> 17.2557032  7.8894098  7.4523847  5.9148024  4.8465381  4.7407573  4.6663889 
#>        V39        V10        V23        V52        V26         V2         V9 
#>  4.2157521  4.1738506  4.1452083  3.2069852  2.8902310  2.8463709  2.7267635 
#>        V53        V31        V40        V56        V20        V46        V50 
#>  2.3104267  1.7584407  1.7146518  1.6825817  1.5758489  1.4630150  1.4222267 
#>        V14         V7         V6        V37        V34        V36        V24 
#>  1.3933895  1.3766516  1.2255193  1.2213645  1.1648590  1.0925790  1.0901473 
#>        V15        V45         V3         V1        V13        V16        V18 
#>  1.0454729  0.8283966  0.6632826  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V19        V21        V22        V25        V27        V28        V29 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V30        V32        V33        V35        V38        V41        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V43        V44        V47        V48        V54        V55        V57 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58        V59        V60         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913