Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 M (0.50359712 0.49640288)  
#>   2) V11>=0.19825 79 17 M (0.78481013 0.21518987)  
#>     4) V16< 0.6636 64  4 M (0.93750000 0.06250000) *
#>     5) V16>=0.6636 15  2 R (0.13333333 0.86666667) *
#>   3) V11< 0.19825 60  8 R (0.13333333 0.86666667)  
#>     6) V1>=0.0392 8  3 M (0.62500000 0.37500000) *
#>     7) V1< 0.0392 52  3 R (0.05769231 0.94230769) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V36< 0.4388 83 24 M (0.71084337 0.28915663)  
#>      4) V46>=0.06695 58  5 M (0.91379310 0.08620690) *
#>      5) V46< 0.06695 25  6 R (0.24000000 0.76000000)  
#>       10) V12>=0.33085 7  1 M (0.85714286 0.14285714) *
#>       11) V12< 0.33085 18  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.4388 56  7 R (0.12500000 0.87500000)  
#>      6) V45>=0.3483 7  2 M (0.71428571 0.28571429) *
#>      7) V45< 0.3483 49  2 R (0.04081633 0.95918367) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V12>=0.20075 83 30 M (0.63855422 0.36144578)  
#>      4) V52>=0.00935 43  6 M (0.86046512 0.13953488)  
#>        8) V39< 0.44435 36  1 M (0.97222222 0.02777778) *
#>        9) V39>=0.44435 7  2 R (0.28571429 0.71428571) *
#>      5) V52< 0.00935 40 16 R (0.40000000 0.60000000)  
#>       10) V23>=0.842 10  0 M (1.00000000 0.00000000) *
#>       11) V23< 0.842 30  6 R (0.20000000 0.80000000)  
#>         22) V27>=0.79985 7  2 M (0.71428571 0.28571429) *
#>         23) V27< 0.79985 23  1 R (0.04347826 0.95652174) *
#>    3) V12< 0.20075 56  9 R (0.16071429 0.83928571)  
#>      6) V20>=0.782 12  4 M (0.66666667 0.33333333) *
#>      7) V20< 0.782 44  1 R (0.02272727 0.97727273) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V9>=0.11655 98 29 M (0.70408163 0.29591837)  
#>      4) V49>=0.02385 80 15 M (0.81250000 0.18750000)  
#>        8) V3>=0.0161 67  6 M (0.91044776 0.08955224)  
#>         16) V36< 0.56485 58  1 M (0.98275862 0.01724138) *
#>         17) V36>=0.56485 9  4 R (0.44444444 0.55555556) *
#>        9) V3< 0.0161 13  4 R (0.30769231 0.69230769) *
#>      5) V49< 0.02385 18  4 R (0.22222222 0.77777778) *
#>    3) V9< 0.11655 41  1 R (0.02439024 0.97560976) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V9>=0.16295 65 14 M (0.78461538 0.21538462)  
#>      4) V14< 0.4046 46  1 M (0.97826087 0.02173913) *
#>      5) V14>=0.4046 19  6 R (0.31578947 0.68421053) *
#>    3) V9< 0.16295 74 20 R (0.27027027 0.72972973)  
#>      6) V4>=0.0606 19  5 M (0.73684211 0.26315789) *
#>      7) V4< 0.0606 55  6 R (0.10909091 0.89090909)  
#>       14) V2>=0.0457 7  3 M (0.57142857 0.42857143) *
#>       15) V2< 0.0457 48  2 R (0.04166667 0.95833333) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 M (0.52517986 0.47482014)  
#>   2) V52>=0.00905 90 26 M (0.71111111 0.28888889)  
#>     4) V11>=0.1117 75 12 M (0.84000000 0.16000000)  
#>       8) V33< 0.73755 68  7 M (0.89705882 0.10294118) *
#>       9) V33>=0.73755 7  2 R (0.28571429 0.71428571) *
#>     5) V11< 0.1117 15  1 R (0.06666667 0.93333333) *
#>   3) V52< 0.00905 49  9 R (0.18367347 0.81632653)  
#>     6) V45>=0.31975 7  1 M (0.85714286 0.14285714) *
#>     7) V45< 0.31975 42  3 R (0.07142857 0.92857143) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.5107914 0.4892086)  
#>    2) V10>=0.1663 80 25 M (0.6875000 0.3125000)  
#>      4) V17< 0.703 57  7 M (0.8771930 0.1228070)  
#>        8) V6< 0.1947 50  2 M (0.9600000 0.0400000) *
#>        9) V6>=0.1947 7  2 R (0.2857143 0.7142857) *
#>      5) V17>=0.703 23  5 R (0.2173913 0.7826087)  
#>       10) V32< 0.2922 8  3 M (0.6250000 0.3750000) *
#>       11) V32>=0.2922 15  0 R (0.0000000 1.0000000) *
#>    3) V10< 0.1663 59 16 R (0.2711864 0.7288136)  
#>      6) V5>=0.08545 9  1 M (0.8888889 0.1111111) *
#>      7) V5< 0.08545 50  8 R (0.1600000 0.8400000)  
#>       14) V45>=0.1793 13  5 M (0.6153846 0.3846154) *
#>       15) V45< 0.1793 37  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.43165468 0.56834532)  
#>    2) V9>=0.13175 92 38 M (0.58695652 0.41304348)  
#>      4) V57< 0.01295 71 18 M (0.74647887 0.25352113)  
#>        8) V21>=0.56055 48  5 M (0.89583333 0.10416667)  
#>         16) V51>=0.01235 40  0 M (1.00000000 0.00000000) *
#>         17) V51< 0.01235 8  3 R (0.37500000 0.62500000) *
#>        9) V21< 0.56055 23 10 R (0.43478261 0.56521739)  
#>         18) V27>=0.7776 10  0 M (1.00000000 0.00000000) *
#>         19) V27< 0.7776 13  0 R (0.00000000 1.00000000) *
#>      5) V57>=0.01295 21  1 R (0.04761905 0.95238095) *
#>    3) V9< 0.13175 47  6 R (0.12765957 0.87234043) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 49 M (0.64748201 0.35251799)  
#>   2) V9>=0.1073 105 23 M (0.78095238 0.21904762)  
#>     4) V6< 0.21915 96 14 M (0.85416667 0.14583333)  
#>       8) V29< 0.93875 89  8 M (0.91011236 0.08988764) *
#>       9) V29>=0.93875 7  1 R (0.14285714 0.85714286) *
#>     5) V6>=0.21915 9  0 R (0.00000000 1.00000000) *
#>   3) V9< 0.1073 34  8 R (0.23529412 0.76470588)  
#>     6) V23>=0.9699 7  0 M (1.00000000 0.00000000) *
#>     7) V23< 0.9699 27  1 R (0.03703704 0.96296296) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V43>=0.23825 72 17 M (0.76388889 0.23611111)  
#>      4) V5>=0.04005 57  5 M (0.91228070 0.08771930)  
#>        8) V35< 0.76075 50  1 M (0.98000000 0.02000000) *
#>        9) V35>=0.76075 7  3 R (0.42857143 0.57142857) *
#>      5) V5< 0.04005 15  3 R (0.20000000 0.80000000) *
#>    3) V43< 0.23825 67 17 R (0.25373134 0.74626866)  
#>      6) V52>=0.00935 36 15 R (0.41666667 0.58333333)  
#>       12) V8< 0.19715 22  8 M (0.63636364 0.36363636)  
#>         24) V13>=0.176 15  2 M (0.86666667 0.13333333) *
#>         25) V13< 0.176 7  1 R (0.14285714 0.85714286) *
#>       13) V8>=0.19715 14  1 R (0.07142857 0.92857143) *
#>      7) V52< 0.00935 31  2 R (0.06451613 0.93548387) *
#> 
#> 
#> $weights
#>  [1] 0.8355657 0.7497369 0.9704602 0.6595794 0.6600047 0.6851856 0.7329966
#>  [8] 0.8419177 0.6947887 0.8503752
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0877900 5.5928208
#>   [2,] 1.4093164 6.2712944
#>   [3,] 1.4349226 6.2456881
#>   [4,] 1.4827336 6.1978772
#>   [5,] 2.2800914 5.4005193
#>   [6,] 2.1041050 5.5765057
#>   [7,] 0.7497369 6.9308738
#>   [8,] 0.0000000 7.6806107
#>   [9,] 2.2949008 5.3857099
#>  [10,] 1.8208354 5.8597754
#>  [11,] 1.7034568 5.9771539
#>  [12,] 1.4827336 6.1978772
#>  [13,] 2.2285670 5.4520437
#>  [14,] 0.0000000 7.6806107
#>  [15,] 2.8274986 4.8531121
#>  [16,] 1.4955704 6.1850403
#>  [17,] 3.6782991 4.0023117
#>  [18,] 0.7329966 6.9476141
#>  [19,] 0.6600047 7.0206061
#>  [20,] 0.7329966 6.9476141
#>  [21,] 1.3930013 6.2876095
#>  [22,] 1.6556458 6.0249649
#>  [23,] 1.4349226 6.2456881
#>  [24,] 1.8208354 5.8597754
#>  [25,] 1.4955704 6.1850403
#>  [26,] 2.0143728 5.6662380
#>  [27,] 2.2155401 5.4650706
#>  [28,] 0.0000000 7.6806107
#>  [29,] 0.8503752 6.8302355
#>  [30,] 0.6851856 6.9954251
#>  [31,] 2.0781869 5.6024238
#>  [32,] 0.7329966 6.9476141
#>  [33,] 1.6859409 5.9946698
#>  [34,] 2.2285670 5.4520437
#>  [35,] 2.0399790 5.6406317
#>  [36,] 1.4181822 6.2624285
#>  [37,] 0.6851856 6.9954251
#>  [38,] 1.3930013 6.2876095
#>  [39,] 0.0000000 7.6806107
#>  [40,] 0.0000000 7.6806107
#>  [41,] 0.7497369 6.9308738
#>  [42,] 2.2864433 5.3941674
#>  [43,] 1.7201971 5.9604136
#>  [44,] 0.6947887 6.9858220
#>  [45,] 0.6947887 6.9858220
#>  [46,] 1.4277853 6.2528254
#>  [47,] 0.0000000 7.6806107
#>  [48,] 0.0000000 7.6806107
#>  [49,] 0.0000000 7.6806107
#>  [50,] 0.0000000 7.6806107
#>  [51,] 2.1679192 5.5126915
#>  [52,] 0.7497369 6.9308738
#>  [53,] 1.4827336 6.1978772
#>  [54,] 0.9704602 6.7101505
#>  [55,] 1.6652489 6.0153619
#>  [56,] 2.5071665 5.1734442
#>  [57,] 2.3248283 5.3557825
#>  [58,] 0.7497369 6.9308738
#>  [59,] 3.1001715 4.5804393
#>  [60,] 1.4445256 6.2360851
#>  [61,] 2.7641097 4.9165010
#>  [62,] 2.1129709 5.5676398
#>  [63,] 2.1129709 5.5676398
#>  [64,] 2.1129709 5.5676398
#>  [65,] 2.1129709 5.5676398
#>  [66,] 1.4277853 6.2528254
#>  [67,] 1.5271033 6.1535074
#>  [68,] 1.6556458 6.0249649
#>  [69,] 1.7201971 5.9604136
#>  [70,] 0.8355657 6.8450450
#>  [71,] 1.3547934 6.3258174
#>  [72,] 0.6947887 6.9858220
#>  [73,] 6.0889561 1.5916546
#>  [74,] 6.8386931 0.8419177
#>  [75,] 5.4520437 2.2285670
#>  [76,] 6.1120484 1.5685623
#>  [77,] 5.2701307 2.4104800
#>  [78,] 5.4843250 2.1962858
#>  [79,] 5.4939280 2.1866827
#>  [80,] 5.3435479 2.3370628
#>  [81,] 5.3435479 2.3370628
#>  [82,] 6.0889561 1.5916546
#>  [83,] 6.8386931 0.8419177
#>  [84,] 7.0210313 0.6595794
#>  [85,] 7.6806107 0.0000000
#>  [86,] 5.6024238 2.0781869
#>  [87,] 6.1786884 1.5019223
#>  [88,] 5.5106515 2.1699593
#>  [89,] 6.1702309 1.5103799
#>  [90,] 6.1850403 1.4955704
#>  [91,] 4.8402753 2.8403355
#>  [92,] 5.3346651 2.3459456
#>  [93,] 5.4902516 2.1903591
#>  [94,] 4.5790015 3.1016092
#>  [95,] 5.2001960 2.4804148
#>  [96,] 5.8597754 1.8208354
#>  [97,] 6.1791137 1.5014971
#>  [98,] 5.9946698 1.6859409
#>  [99,] 6.8386931 0.8419177
#> [100,] 6.1439044 1.5367064
#> [101,] 6.8386931 0.8419177
#> [102,] 6.1786884 1.5019223
#> [103,] 7.0206061 0.6600047
#> [104,] 4.5853535 3.0952573
#> [105,] 6.0804986 1.6001121
#> [106,] 6.0031273 1.6774834
#> [107,] 4.6499048 3.0307060
#> [108,] 7.0210313 0.6595794
#> [109,] 5.3653855 2.3152252
#> [110,] 7.0210313 0.6595794
#> [111,] 6.0505711 1.6300396
#> [112,] 5.2082282 2.4723825
#> [113,] 4.7441911 2.9364197
#> [114,] 3.9739451 3.7066656
#> [115,] 7.0210313 0.6595794
#> [116,] 7.6806107 0.0000000
#> [117,] 5.5191090 2.1615017
#> [118,] 6.1702309 1.5103799
#> [119,] 5.4854705 2.1951402
#> [120,] 6.1850403 1.4955704
#> [121,] 5.4838997 2.1967110
#> [122,] 6.7101505 0.9704602
#> [123,] 6.0501459 1.6304649
#> [124,] 5.9771539 1.7034568
#> [125,] 6.7101505 0.9704602
#> [126,] 5.8682329 1.8123778
#> [127,] 6.3262426 1.3543681
#> [128,] 6.1786884 1.5019223
#> [129,] 6.1450499 1.5355608
#> [130,] 6.3006364 1.3799743
#> [131,] 7.6806107 0.0000000
#> [132,] 7.0206061 0.6600047
#> [133,] 7.6806107 0.0000000
#> [134,] 7.6806107 0.0000000
#> [135,] 7.6806107 0.0000000
#> [136,] 6.8302355 0.8503752
#> [137,] 6.1786884 1.5019223
#> [138,] 5.3031322 2.3774785
#> [139,] 6.1706561 1.5099546
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.27182604 0.72817396
#>   [2,] 0.18349014 0.81650986
#>   [3,] 0.18682402 0.81317598
#>   [4,] 0.19304891 0.80695109
#>   [5,] 0.29686329 0.70313671
#>   [6,] 0.27395023 0.72604977
#>   [7,] 0.09761424 0.90238576
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.29879145 0.70120855
#>  [10,] 0.23706909 0.76293091
#>  [11,] 0.22178663 0.77821337
#>  [12,] 0.19304891 0.80695109
#>  [13,] 0.29015492 0.70984508
#>  [14,] 0.00000000 1.00000000
#>  [15,] 0.36813461 0.63186539
#>  [16,] 0.19472025 0.80527975
#>  [17,] 0.47890711 0.52109289
#>  [18,] 0.09543468 0.90456532
#>  [19,] 0.08593128 0.91406872
#>  [20,] 0.09543468 0.90456532
#>  [21,] 0.18136595 0.81863405
#>  [22,] 0.21556174 0.78443826
#>  [23,] 0.18682402 0.81317598
#>  [24,] 0.23706909 0.76293091
#>  [25,] 0.19472025 0.80527975
#>  [26,] 0.26226727 0.73773273
#>  [27,] 0.28845884 0.71154116
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.11071713 0.88928287
#>  [30,] 0.08920979 0.91079021
#>  [31,] 0.27057574 0.72942426
#>  [32,] 0.09543468 0.90456532
#>  [33,] 0.21950610 0.78049390
#>  [34,] 0.29015492 0.70984508
#>  [35,] 0.26560115 0.73439885
#>  [36,] 0.18464446 0.81535554
#>  [37,] 0.08920979 0.91079021
#>  [38,] 0.18136595 0.81863405
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.09761424 0.90238576
#>  [42,] 0.29769030 0.70230970
#>  [43,] 0.22396619 0.77603381
#>  [44,] 0.09046008 0.90953992
#>  [45,] 0.09046008 0.90953992
#>  [46,] 0.18589476 0.81410524
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.00000000 1.00000000
#>  [51,] 0.28225870 0.71774130
#>  [52,] 0.09761424 0.90238576
#>  [53,] 0.19304891 0.80695109
#>  [54,] 0.12635195 0.87364805
#>  [55,] 0.21681204 0.78318796
#>  [56,] 0.32642802 0.67357198
#>  [57,] 0.30268794 0.69731206
#>  [58,] 0.09761424 0.90238576
#>  [59,] 0.40363606 0.59636394
#>  [60,] 0.18807432 0.81192568
#>  [61,] 0.35988150 0.64011850
#>  [62,] 0.27510455 0.72489545
#>  [63,] 0.27510455 0.72489545
#>  [64,] 0.27510455 0.72489545
#>  [65,] 0.27510455 0.72489545
#>  [66,] 0.18589476 0.81410524
#>  [67,] 0.19882576 0.80117424
#>  [68,] 0.21556174 0.78443826
#>  [69,] 0.22396619 0.77603381
#>  [70,] 0.10878897 0.89121103
#>  [71,] 0.17639136 0.82360864
#>  [72,] 0.09046008 0.90953992
#>  [73,] 0.79276979 0.20723021
#>  [74,] 0.89038402 0.10961598
#>  [75,] 0.70984508 0.29015492
#>  [76,] 0.79577635 0.20422365
#>  [77,] 0.68616037 0.31383963
#>  [78,] 0.71404803 0.28595197
#>  [79,] 0.71529833 0.28470167
#>  [80,] 0.69571914 0.30428086
#>  [81,] 0.69571914 0.30428086
#>  [82,] 0.79276979 0.20723021
#>  [83,] 0.89038402 0.10961598
#>  [84,] 0.91412409 0.08587591
#>  [85,] 1.00000000 0.00000000
#>  [86,] 0.72942426 0.27057574
#>  [87,] 0.80445275 0.19554725
#>  [88,] 0.71747569 0.28252431
#>  [89,] 0.80335159 0.19664841
#>  [90,] 0.80527975 0.19472025
#>  [91,] 0.63019406 0.36980594
#>  [92,] 0.69456262 0.30543738
#>  [93,] 0.71481967 0.28518033
#>  [94,] 0.59617675 0.40382325
#>  [95,] 0.67705501 0.32294499
#>  [96,] 0.76293091 0.23706909
#>  [97,] 0.80450812 0.19549188
#>  [98,] 0.78049390 0.21950610
#>  [99,] 0.89038402 0.10961598
#> [100,] 0.79992394 0.20007606
#> [101,] 0.89038402 0.10961598
#> [102,] 0.80445275 0.19554725
#> [103,] 0.91406872 0.08593128
#> [104,] 0.59700376 0.40299624
#> [105,] 0.79166863 0.20833137
#> [106,] 0.78159505 0.21840495
#> [107,] 0.60540820 0.39459180
#> [108,] 0.91412409 0.08587591
#> [109,] 0.69856235 0.30143765
#> [110,] 0.91412409 0.08587591
#> [111,] 0.78777214 0.21222786
#> [112,] 0.67810079 0.32189921
#> [113,] 0.61768409 0.38231591
#> [114,] 0.51739963 0.48260037
#> [115,] 0.91412409 0.08587591
#> [116,] 1.00000000 0.00000000
#> [117,] 0.71857684 0.28142316
#> [118,] 0.80335159 0.19664841
#> [119,] 0.71419718 0.28580282
#> [120,] 0.80527975 0.19472025
#> [121,] 0.71399266 0.28600734
#> [122,] 0.87364805 0.12635195
#> [123,] 0.78771677 0.21228323
#> [124,] 0.77821337 0.22178663
#> [125,] 0.87364805 0.12635195
#> [126,] 0.76403207 0.23596793
#> [127,] 0.82366401 0.17633599
#> [128,] 0.80445275 0.19554725
#> [129,] 0.80007308 0.19992692
#> [130,] 0.82033013 0.17966987
#> [131,] 1.00000000 0.00000000
#> [132,] 0.91406872 0.08593128
#> [133,] 1.00000000 0.00000000
#> [134,] 1.00000000 0.00000000
#> [135,] 1.00000000 0.00000000
#> [136,] 0.88928287 0.11071713
#> [137,] 0.80445275 0.19554725
#> [138,] 0.69045710 0.30954290
#> [139,] 0.80340696 0.19659304
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.9896194  2.2896991  9.1374374  5.4063149  1.1285841  2.0672238  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  3.4851786  2.7755537  0.0000000  0.0000000  0.6006109  2.0138251  1.4770184 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  4.3733637  0.0000000  0.0000000  0.0000000  3.7696691  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  1.4088261  1.3847957  0.0000000  0.0000000  0.7929094  0.8626491  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.8427338  5.3573963  0.0000000  0.0000000  1.4226860  1.9494622  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  4.0768041  0.0000000  3.8699457  3.1561678  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  0.0000000  1.7923570  4.2957822  0.0000000  1.1637147  6.4053877  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  3.5371031  0.0000000  0.0000000  3.3001176 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  1.2324349 13.6346285 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 67 72 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>         V9        V11        V52        V12        V36        V23         V5 
#> 13.6346285  9.1374374  6.4053877  5.4063149  5.3573963  4.3733637  4.2957822 
#>        V43        V45        V27        V57        V16         V6        V46 
#>  4.0768041  3.8699457  3.7696691  3.5371031  3.4851786  3.3001176  3.1561678 
#>        V17        V10        V14        V20         V4        V49        V21 
#>  2.7755537  2.2896991  2.0672238  2.0138251  1.9494622  1.7923570  1.4770184 
#>        V39        V29         V3         V8        V51        V13         V1 
#>  1.4226860  1.4088261  1.3847957  1.2324349  1.1637147  1.1285841  0.9896194 
#>        V33        V35        V32         V2        V15        V18        V19 
#>  0.8626491  0.8427338  0.7929094  0.6006109  0.0000000  0.0000000  0.0000000 
#>        V22        V24        V25        V26        V28        V30        V31 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V34        V37        V38        V40        V41        V42        V44 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V47        V48        V50        V53        V54        V55        V56 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58        V59        V60         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913