Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 M (0.61870504 0.38129496)  
#>    2) V9>=0.10195 100 21 M (0.79000000 0.21000000)  
#>      4) V12>=0.12175 92 13 M (0.85869565 0.14130435)  
#>        8) V27>=0.82225 62  1 M (0.98387097 0.01612903) *
#>        9) V27< 0.82225 30 12 M (0.60000000 0.40000000)  
#>         18) V50>=0.0177 21  3 M (0.85714286 0.14285714) *
#>         19) V50< 0.0177 9  0 R (0.00000000 1.00000000) *
#>      5) V12< 0.12175 8  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.10195 39  7 R (0.17948718 0.82051282)  
#>      6) V31< 0.4113 13  6 M (0.53846154 0.46153846) *
#>      7) V31>=0.4113 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 M (0.51079137 0.48920863)  
#>   2) V11>=0.19795 84 23 M (0.72619048 0.27380952)  
#>     4) V41< 0.41295 65 10 M (0.84615385 0.15384615)  
#>       8) V52>=0.0086 54  4 M (0.92592593 0.07407407) *
#>       9) V52< 0.0086 11  5 R (0.45454545 0.54545455) *
#>     5) V41>=0.41295 19  6 R (0.31578947 0.68421053) *
#>   3) V11< 0.19795 55 10 R (0.18181818 0.81818182)  
#>     6) V7>=0.13185 11  2 M (0.81818182 0.18181818) *
#>     7) V7< 0.13185 44  1 R (0.02272727 0.97727273) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V9>=0.11575 95 24 M (0.74736842 0.25263158)  
#>     4) V49>=0.0236 82 13 M (0.84146341 0.15853659)  
#>       8) V16< 0.6636 65  1 M (0.98461538 0.01538462) *
#>       9) V16>=0.6636 17  5 R (0.29411765 0.70588235) *
#>     5) V49< 0.0236 13  2 R (0.15384615 0.84615385) *
#>   3) V9< 0.11575 44  6 R (0.13636364 0.86363636)  
#>     6) V53>=0.01495 7  3 M (0.57142857 0.42857143) *
#>     7) V53< 0.01495 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 M (0.54676259 0.45323741)  
#>   2) V11>=0.19795 78 16 M (0.79487179 0.20512821)  
#>     4) V48>=0.0444 65  7 M (0.89230769 0.10769231)  
#>       8) V18< 0.9243 58  1 M (0.98275862 0.01724138) *
#>       9) V18>=0.9243 7  1 R (0.14285714 0.85714286) *
#>     5) V48< 0.0444 13  4 R (0.30769231 0.69230769) *
#>   3) V11< 0.19795 61 14 R (0.22950820 0.77049180)  
#>     6) V22>=0.99195 9  0 M (1.00000000 0.00000000) *
#>     7) V22< 0.99195 52  5 R (0.09615385 0.90384615) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V12>=0.15395 93 30 M (0.67741935 0.32258065)  
#>      4) V48>=0.07185 56  6 M (0.89285714 0.10714286) *
#>      5) V48< 0.07185 37 13 R (0.35135135 0.64864865)  
#>       10) V27>=0.90665 11  1 M (0.90909091 0.09090909) *
#>       11) V27< 0.90665 26  3 R (0.11538462 0.88461538) *
#>    3) V12< 0.15395 46  8 R (0.17391304 0.82608696)  
#>      6) V20>=0.8328 12  5 M (0.58333333 0.41666667) *
#>      7) V20< 0.8328 34  1 R (0.02941176 0.97058824) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V9>=0.15375 62 17 M (0.72580645 0.27419355)  
#>      4) V15< 0.5607 41  3 M (0.92682927 0.07317073) *
#>      5) V15>=0.5607 21  7 R (0.33333333 0.66666667)  
#>       10) V30< 0.44475 7  0 M (1.00000000 0.00000000) *
#>       11) V30>=0.44475 14  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.15375 77 17 R (0.22077922 0.77922078)  
#>      6) V28>=0.985 10  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.985 67  7 R (0.10447761 0.89552239)  
#>       14) V4>=0.0602 9  4 M (0.55555556 0.44444444) *
#>       15) V4< 0.0602 58  2 R (0.03448276 0.96551724) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.5467626 0.4532374)  
#>    2) V33< 0.58425 108 36 M (0.6666667 0.3333333)  
#>      4) V12>=0.1926 78 15 M (0.8076923 0.1923077)  
#>        8) V17< 0.92735 71  9 M (0.8732394 0.1267606)  
#>         16) V10>=0.08865 64  4 M (0.9375000 0.0625000) *
#>         17) V10< 0.08865 7  2 R (0.2857143 0.7142857) *
#>        9) V17>=0.92735 7  1 R (0.1428571 0.8571429) *
#>      5) V12< 0.1926 30  9 R (0.3000000 0.7000000)  
#>       10) V25< 0.65375 14  5 M (0.6428571 0.3571429) *
#>       11) V25>=0.65375 16  0 R (0.0000000 1.0000000) *
#>    3) V33>=0.58425 31  4 R (0.1290323 0.8709677)  
#>      6) V14>=0.27205 7  3 M (0.5714286 0.4285714) *
#>      7) V14< 0.27205 24  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V44>=0.29145 24  1 M (0.95833333 0.04166667) *
#>    3) V44< 0.29145 115 38 R (0.33043478 0.66956522)  
#>      6) V31< 0.4623 52 20 M (0.61538462 0.38461538)  
#>       12) V32>=0.08755 40  8 M (0.80000000 0.20000000)  
#>         24) V7>=0.1079 30  1 M (0.96666667 0.03333333) *
#>         25) V7< 0.1079 10  3 R (0.30000000 0.70000000) *
#>       13) V32< 0.08755 12  0 R (0.00000000 1.00000000) *
#>      7) V31>=0.4623 63  6 R (0.09523810 0.90476190) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V55< 0.00685 63 16 M (0.74603175 0.25396825)  
#>      4) V12>=0.1694 45  2 M (0.95555556 0.04444444) *
#>      5) V12< 0.1694 18  4 R (0.22222222 0.77777778) *
#>    3) V55>=0.00685 76 24 R (0.31578947 0.68421053)  
#>      6) V52>=0.0238 11  0 M (1.00000000 0.00000000) *
#>      7) V52< 0.0238 65 13 R (0.20000000 0.80000000)  
#>       14) V23>=0.80715 14  4 M (0.71428571 0.28571429) *
#>       15) V23< 0.80715 51  3 R (0.05882353 0.94117647) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V27>=0.8191 47  7 M (0.85106383 0.14893617)  
#>      4) V12>=0.222 36  1 M (0.97222222 0.02777778) *
#>      5) V12< 0.222 11  5 R (0.45454545 0.54545455) *
#>    3) V27< 0.8191 92 32 R (0.34782609 0.65217391)  
#>      6) V16>=0.3734 55 24 M (0.56363636 0.43636364)  
#>       12) V15< 0.4746 24  1 M (0.95833333 0.04166667) *
#>       13) V15>=0.4746 31  8 R (0.25806452 0.74193548)  
#>         26) V26< 0.5635 7  0 M (1.00000000 0.00000000) *
#>         27) V26>=0.5635 24  1 R (0.04166667 0.95833333) *
#>      7) V16< 0.3734 37  1 R (0.02702703 0.97297297) *
#> 
#> 
#> $weights
#>  [1] 0.8090480 0.6617672 0.6209828 0.7268909 0.6524534 1.0417664 0.6174296
#>  [8] 0.7220292 0.7309319 0.8494180
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.8461919 4.5865255
#>   [2,] 1.3744826 6.0582348
#>   [3,] 2.5086141 4.9241034
#>   [4,] 1.4578228 5.9748947
#>   [5,] 3.0111107 4.4216067
#>   [6,] 2.8968784 4.5358390
#>   [7,] 2.0789311 5.3537864
#>   [8,] 0.0000000 7.4327174
#>   [9,] 1.2384124 6.1943050
#>  [10,] 2.1924333 5.2402841
#>  [11,] 2.2668708 5.1658467
#>  [12,] 2.3209633 5.1117542
#>  [13,] 3.0554483 4.3772691
#>  [14,] 0.0000000 7.4327174
#>  [15,] 0.0000000 7.4327174
#>  [16,] 0.8494180 6.5832994
#>  [17,] 1.6584660 5.7742515
#>  [18,] 1.4615014 5.9712160
#>  [19,] 2.6311115 4.8016059
#>  [20,] 2.3152026 5.1175148
#>  [21,] 0.6209828 6.8117347
#>  [22,] 0.0000000 7.4327174
#>  [23,] 0.0000000 7.4327174
#>  [24,] 1.6591961 5.7735214
#>  [25,] 0.6174296 6.8152878
#>  [26,] 0.0000000 7.4327174
#>  [27,] 2.8905295 4.5421879
#>  [28,] 2.6222088 4.8105086
#>  [29,] 2.5728436 4.8598738
#>  [30,] 2.3860869 5.0466305
#>  [31,] 1.3793443 6.0533731
#>  [32,] 1.9604416 5.4722758
#>  [33,] 0.0000000 7.4327174
#>  [34,] 0.6524534 6.7802640
#>  [35,] 0.0000000 7.4327174
#>  [36,] 1.4300308 6.0026867
#>  [37,] 2.0008150 5.4319024
#>  [38,] 0.0000000 7.4327174
#>  [39,] 1.0417664 6.3909510
#>  [40,] 1.0417664 6.3909510
#>  [41,] 0.6174296 6.8152878
#>  [42,] 0.6524534 6.7802640
#>  [43,] 1.2734362 6.1592812
#>  [44,] 0.7220292 6.7106882
#>  [45,] 0.0000000 7.4327174
#>  [46,] 2.1520600 5.2806575
#>  [47,] 0.8090480 6.6236694
#>  [48,] 0.8090480 6.6236694
#>  [49,] 0.8090480 6.6236694
#>  [50,] 1.3744826 6.0582348
#>  [51,] 2.0411115 5.3916059
#>  [52,] 1.4704008 5.9623167
#>  [53,] 2.3072408 5.1254767
#>  [54,] 2.9157677 4.5169497
#>  [55,] 2.1232686 5.3094488
#>  [56,] 1.4615014 5.9712160
#>  [57,] 0.6524534 6.7802640
#>  [58,] 2.1195900 5.3131275
#>  [59,] 2.2579681 5.1747494
#>  [60,] 0.7268909 6.7058266
#>  [61,] 1.4264776 6.0062398
#>  [62,] 1.6591961 5.7735214
#>  [63,] 1.9001796 5.5325378
#>  [64,] 1.6591961 5.7735214
#>  [65,] 5.8564086 1.5763089
#>  [66,] 7.4327174 0.0000000
#>  [67,] 5.2753079 2.1574095
#>  [68,] 5.4323903 2.0003271
#>  [69,] 6.0897055 1.3430120
#>  [70,] 5.4372520 1.9954654
#>  [71,] 5.3220302 2.1106873
#>  [72,] 5.4359435 1.9967739
#>  [73,] 5.3574650 2.0752524
#>  [74,] 5.4314914 2.0012260
#>  [75,] 5.0116067 2.4211107
#>  [76,] 5.8612702 1.5714472
#>  [77,] 5.3220302 2.1106873
#>  [78,] 7.4327174 0.0000000
#>  [79,] 6.7709502 0.6617672
#>  [80,] 7.4327174 0.0000000
#>  [81,] 6.0848438 1.3478736
#>  [82,] 7.4327174 0.0000000
#>  [83,] 6.0026867 1.4300308
#>  [84,] 4.6831198 2.7495976
#>  [85,] 5.3098632 2.1228542
#>  [86,] 5.5418516 1.8908659
#>  [87,] 5.0433195 2.3893979
#>  [88,] 6.7709502 0.6617672
#>  [89,] 6.6236694 0.8090480
#>  [90,] 6.7106882 0.7220292
#>  [91,] 6.0883969 1.3443205
#>  [92,] 6.7058266 0.7268909
#>  [93,] 5.4225887 2.0101287
#>  [94,] 5.9215322 1.5111852
#>  [95,] 6.5832994 0.8494180
#>  [96,] 5.9308460 1.5018714
#>  [97,] 6.0489210 1.3837964
#>  [98,] 4.6193013 2.8134161
#>  [99,] 5.9797563 1.4529611
#> [100,] 6.3909510 1.0417664
#> [101,] 5.2717548 2.1609627
#> [102,] 5.1658467 2.2668708
#> [103,] 6.1535206 1.2791968
#> [104,] 6.1535206 1.2791968
#> [105,] 6.1535206 1.2791968
#> [106,] 5.4225887 2.0101287
#> [107,] 4.8797658 2.5529516
#> [108,] 5.1995030 2.2332144
#> [109,] 3.1902733 4.2424442
#> [110,] 5.6640601 1.7686573
#> [111,] 5.0430774 2.3896401
#> [112,] 6.0932586 1.3394588
#> [113,] 6.7709502 0.6617672
#> [114,] 6.1499675 1.2827500
#> [115,] 6.0440594 1.3886581
#> [116,] 4.2802637 3.1524537
#> [117,] 6.7017855 0.7309319
#> [118,] 6.0400183 1.3926991
#> [119,] 6.7017855 0.7309319
#> [120,] 6.7017855 0.7309319
#> [121,] 6.7017855 0.7309319
#> [122,] 5.9215322 1.5111852
#> [123,] 6.7106882 0.7220292
#> [124,] 6.7106882 0.7220292
#> [125,] 6.7106882 0.7220292
#> [126,] 5.3587736 2.0739439
#> [127,] 6.0489210 1.3837964
#> [128,] 5.3179891 2.1147283
#> [129,] 6.7106882 0.7220292
#> [130,] 6.7106882 0.7220292
#> [131,] 6.7106882 0.7220292
#> [132,] 7.4327174 0.0000000
#> [133,] 7.4327174 0.0000000
#> [134,] 6.7106882 0.7220292
#> [135,] 6.7106882 0.7220292
#> [136,] 5.4279383 2.0047792
#> [137,] 6.7106882 0.7220292
#> [138,] 6.7106882 0.7220292
#> [139,] 5.3628146 2.0699028
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.38292750 0.61707250
#>   [2,] 0.18492330 0.81507670
#>   [3,] 0.33750968 0.66249032
#>   [4,] 0.19613591 0.80386409
#>   [5,] 0.40511573 0.59488427
#>   [6,] 0.38974688 0.61025312
#>   [7,] 0.27970000 0.72030000
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.16661637 0.83338363
#>  [10,] 0.29497063 0.70502937
#>  [11,] 0.30498546 0.69501454
#>  [12,] 0.31226308 0.68773692
#>  [13,] 0.41108092 0.58891908
#>  [14,] 0.00000000 1.00000000
#>  [15,] 0.00000000 1.00000000
#>  [16,] 0.11428095 0.88571905
#>  [17,] 0.22313050 0.77686950
#>  [18,] 0.19663083 0.80336917
#>  [19,] 0.35399052 0.64600948
#>  [20,] 0.31148805 0.68851195
#>  [21,] 0.08354721 0.91645279
#>  [22,] 0.00000000 1.00000000
#>  [23,] 0.00000000 1.00000000
#>  [24,] 0.22322873 0.77677127
#>  [25,] 0.08306917 0.91693083
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.38889269 0.61110731
#>  [28,] 0.35279275 0.64720725
#>  [29,] 0.34615114 0.65384886
#>  [30,] 0.32102484 0.67897516
#>  [31,] 0.18557739 0.81442261
#>  [32,] 0.26375839 0.73624161
#>  [33,] 0.00000000 1.00000000
#>  [34,] 0.08778128 0.91221872
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.19239676 0.80760324
#>  [37,] 0.26919024 0.73080976
#>  [38,] 0.00000000 1.00000000
#>  [39,] 0.14015956 0.85984044
#>  [40,] 0.14015956 0.85984044
#>  [41,] 0.08306917 0.91693083
#>  [42,] 0.08778128 0.91221872
#>  [43,] 0.17132848 0.82867152
#>  [44,] 0.09714202 0.90285798
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.28953878 0.71046122
#>  [47,] 0.10884956 0.89115044
#>  [48,] 0.10884956 0.89115044
#>  [49,] 0.10884956 0.89115044
#>  [50,] 0.18492330 0.81507670
#>  [51,] 0.27461174 0.72538826
#>  [52,] 0.19782815 0.80217185
#>  [53,] 0.31041685 0.68958315
#>  [54,] 0.39228825 0.60771175
#>  [55,] 0.28566519 0.71433481
#>  [56,] 0.19663083 0.80336917
#>  [57,] 0.08778128 0.91221872
#>  [58,] 0.28517026 0.71482974
#>  [59,] 0.30378769 0.69621231
#>  [60,] 0.09779611 0.90220389
#>  [61,] 0.19191872 0.80808128
#>  [62,] 0.22322873 0.77677127
#>  [63,] 0.25565073 0.74434927
#>  [64,] 0.22322873 0.77677127
#>  [65,] 0.78792294 0.21207706
#>  [66,] 1.00000000 0.00000000
#>  [67,] 0.70974148 0.29025852
#>  [68,] 0.73087540 0.26912460
#>  [69,] 0.81931077 0.18068923
#>  [70,] 0.73152949 0.26847051
#>  [71,] 0.71602751 0.28397249
#>  [72,] 0.73135344 0.26864656
#>  [73,] 0.72079493 0.27920507
#>  [74,] 0.73075446 0.26924554
#>  [75,] 0.67426305 0.32573695
#>  [76,] 0.78857703 0.21142297
#>  [77,] 0.71602751 0.28397249
#>  [78,] 1.00000000 0.00000000
#>  [79,] 0.91096565 0.08903435
#>  [80,] 1.00000000 0.00000000
#>  [81,] 0.81865668 0.18134332
#>  [82,] 1.00000000 0.00000000
#>  [83,] 0.80760324 0.19239676
#>  [84,] 0.63006833 0.36993167
#>  [85,] 0.71439057 0.28560943
#>  [86,] 0.74560235 0.25439765
#>  [87,] 0.67852970 0.32147030
#>  [88,] 0.91096565 0.08903435
#>  [89,] 0.89115044 0.10884956
#>  [90,] 0.90285798 0.09714202
#>  [91,] 0.81913472 0.18086528
#>  [92,] 0.90220389 0.09779611
#>  [93,] 0.72955668 0.27044332
#>  [94,] 0.79668470 0.20331530
#>  [95,] 0.88571905 0.11428095
#>  [96,] 0.79793777 0.20206223
#>  [97,] 0.81382362 0.18617638
#>  [98,] 0.62148217 0.37851783
#>  [99,] 0.80451818 0.19548182
#> [100,] 0.85984044 0.14015956
#> [101,] 0.70926344 0.29073656
#> [102,] 0.69501454 0.30498546
#> [103,] 0.82789648 0.17210352
#> [104,] 0.82789648 0.17210352
#> [105,] 0.82789648 0.17210352
#> [106,] 0.72955668 0.27044332
#> [107,] 0.65652514 0.34347486
#> [108,] 0.69954268 0.30045732
#> [109,] 0.42922031 0.57077969
#> [110,] 0.76204432 0.23795568
#> [111,] 0.67849712 0.32150288
#> [112,] 0.81978881 0.18021119
#> [113,] 0.91096565 0.08903435
#> [114,] 0.82741844 0.17258156
#> [115,] 0.81316953 0.18683047
#> [116,] 0.57586795 0.42413205
#> [117,] 0.90166021 0.09833979
#> [118,] 0.81262585 0.18737415
#> [119,] 0.90166021 0.09833979
#> [120,] 0.90166021 0.09833979
#> [121,] 0.90166021 0.09833979
#> [122,] 0.79668470 0.20331530
#> [123,] 0.90285798 0.09714202
#> [124,] 0.90285798 0.09714202
#> [125,] 0.90285798 0.09714202
#> [126,] 0.72097098 0.27902902
#> [127,] 0.81382362 0.18617638
#> [128,] 0.71548383 0.28451617
#> [129,] 0.90285798 0.09714202
#> [130,] 0.90285798 0.09714202
#> [131,] 0.90285798 0.09714202
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 0.90285798 0.09714202
#> [135,] 0.90285798 0.09714202
#> [136,] 0.73027642 0.26972358
#> [137,] 0.90285798 0.09714202
#> [138,] 0.90285798 0.09714202
#> [139,] 0.72151466 0.27848534
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  0.8914622  7.7945516 10.7193730  0.0000000  0.5884981  5.7796748 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  5.0630849  1.1304096  1.7251413  0.0000000  0.0000000  0.9563889  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  2.4539198  1.8579891  0.0000000  1.0261936  2.2771326  6.6138498  3.9155559 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  2.6185643  4.0923754  2.2975170  2.3153590  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  1.1869952  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  1.4741025  0.0000000  0.0000000  3.0444869  0.0000000  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.7455095  1.7745686  0.0000000  2.0170068  0.0000000  3.0942570  0.5270148 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  2.5103167  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  3.2810872  0.0000000 13.2276142 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 75 64 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>         V9        V12        V11        V27        V15        V16        V31 
#> 13.2276142 10.7193730  7.7945516  6.6138498  5.7796748  5.0630849  4.0923754 
#>        V28        V48         V7        V52        V44        V30        V55 
#>  3.9155559  3.7455095  3.2810872  3.0942570  3.0444869  2.6185643  2.5103167 
#>        V22        V33        V32        V26        V50        V23        V49 
#>  2.4539198  2.3153590  2.2975170  2.2771326  2.0170068  1.8579891  1.7745686 
#>        V18        V41         V4        V17        V25        V20        V10 
#>  1.7251413  1.4741025  1.1869952  1.1304096  1.0261936  0.9563889  0.8914622 
#>        V14        V53         V1        V13        V19         V2        V21 
#>  0.5884981  0.5270148  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V29         V3        V34        V35        V36        V37 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V38        V39        V40        V42        V43        V45        V46 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V47         V5        V51        V54        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6        V60         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986