Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V9>=0.11595 99 28 M (0.71717172 0.28282828)  
#>     4) V49>=0.02355 81 12 M (0.85185185 0.14814815)  
#>       8) V50>=0.00825 74  6 M (0.91891892 0.08108108) *
#>       9) V50< 0.00825 7  1 R (0.14285714 0.85714286) *
#>     5) V49< 0.02355 18  2 R (0.11111111 0.88888889) *
#>   3) V9< 0.11595 40  6 R (0.15000000 0.85000000)  
#>     6) V54>=0.0166 9  4 M (0.55555556 0.44444444) *
#>     7) V54< 0.0166 31  1 R (0.03225806 0.96774194) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 R (0.49640288 0.50359712)  
#>   2) V52>=0.00935 75 21 M (0.72000000 0.28000000)  
#>     4) V10>=0.06495 67 13 M (0.80597015 0.19402985)  
#>       8) V13>=0.1637 55  5 M (0.90909091 0.09090909) *
#>       9) V13< 0.1637 12  4 R (0.33333333 0.66666667) *
#>     5) V10< 0.06495 8  0 R (0.00000000 1.00000000) *
#>   3) V52< 0.00935 64 15 R (0.23437500 0.76562500)  
#>     6) V44>=0.23105 18  6 M (0.66666667 0.33333333) *
#>     7) V44< 0.23105 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V36< 0.4542 86 30 M (0.65116279 0.34883721)  
#>      4) V10>=0.1506 56 10 M (0.82142857 0.17857143)  
#>        8) V16< 0.6843 45  2 M (0.95555556 0.04444444) *
#>        9) V16>=0.6843 11  3 R (0.27272727 0.72727273) *
#>      5) V10< 0.1506 30 10 R (0.33333333 0.66666667)  
#>       10) V28< 0.5839 10  1 M (0.90000000 0.10000000) *
#>       11) V28>=0.5839 20  1 R (0.05000000 0.95000000) *
#>    3) V36>=0.4542 53 11 R (0.20754717 0.79245283)  
#>      6) V45>=0.252 13  2 M (0.84615385 0.15384615) *
#>      7) V45< 0.252 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 R (0.47482014 0.52517986)  
#>   2) V12>=0.2281 72 19 M (0.73611111 0.26388889)  
#>     4) V15< 0.54245 57  6 M (0.89473684 0.10526316)  
#>       8) V29>=0.49645 48  1 M (0.97916667 0.02083333) *
#>       9) V29< 0.49645 9  4 R (0.44444444 0.55555556) *
#>     5) V15>=0.54245 15  2 R (0.13333333 0.86666667) *
#>   3) V12< 0.2281 67 13 R (0.19402985 0.80597015)  
#>     6) V4>=0.05885 15  6 M (0.60000000 0.40000000) *
#>     7) V4< 0.05885 52  4 R (0.07692308 0.92307692) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V49>=0.01765 111 40 M (0.63963964 0.36036036)  
#>      4) V8>=0.07545 93 25 M (0.73118280 0.26881720)  
#>        8) V37< 0.6068 79 14 M (0.82278481 0.17721519)  
#>         16) V42>=0.08765 72  8 M (0.88888889 0.11111111) *
#>         17) V42< 0.08765 7  1 R (0.14285714 0.85714286) *
#>        9) V37>=0.6068 14  3 R (0.21428571 0.78571429) *
#>      5) V8< 0.07545 18  3 R (0.16666667 0.83333333) *
#>    3) V49< 0.01765 28  2 R (0.07142857 0.92857143) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V37< 0.3766 66 17 M (0.74242424 0.25757576)  
#>      4) V51>=0.01345 36  1 M (0.97222222 0.02777778) *
#>      5) V51< 0.01345 30 14 R (0.46666667 0.53333333)  
#>       10) V40>=0.32035 11  0 M (1.00000000 0.00000000) *
#>       11) V40< 0.32035 19  3 R (0.15789474 0.84210526) *
#>    3) V37>=0.3766 73 21 R (0.28767123 0.71232877)  
#>      6) V31< 0.3638 22  6 M (0.72727273 0.27272727)  
#>       12) V30>=0.34515 15  0 M (1.00000000 0.00000000) *
#>       13) V30< 0.34515 7  1 R (0.14285714 0.85714286) *
#>      7) V31>=0.3638 51  5 R (0.09803922 0.90196078) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V12>=0.1677 92 31 M (0.66304348 0.33695652)  
#>      4) V23>=0.78455 38  1 M (0.97368421 0.02631579) *
#>      5) V23< 0.78455 54 24 R (0.44444444 0.55555556)  
#>       10) V48>=0.1072 16  0 M (1.00000000 0.00000000) *
#>       11) V48< 0.1072 38  8 R (0.21052632 0.78947368)  
#>         22) V55< 0.0067 11  4 M (0.63636364 0.36363636) *
#>         23) V55>=0.0067 27  1 R (0.03703704 0.96296296) *
#>    3) V12< 0.1677 47 11 R (0.23404255 0.76595745)  
#>      6) V25< 0.55335 12  3 M (0.75000000 0.25000000) *
#>      7) V25>=0.55335 35  2 R (0.05714286 0.94285714) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V49>=0.05115 43 10 M (0.76744186 0.23255814)  
#>      4) V39>=0.1107 35  4 M (0.88571429 0.11428571)  
#>        8) V11>=0.2176 28  0 M (1.00000000 0.00000000) *
#>        9) V11< 0.2176 7  3 R (0.42857143 0.57142857) *
#>      5) V39< 0.1107 8  2 R (0.25000000 0.75000000) *
#>    3) V49< 0.05115 96 28 R (0.29166667 0.70833333)  
#>      6) V31< 0.4513 34 12 M (0.64705882 0.35294118)  
#>       12) V40>=0.1397 26  4 M (0.84615385 0.15384615) *
#>       13) V40< 0.1397 8  0 R (0.00000000 1.00000000) *
#>      7) V31>=0.4513 62  6 R (0.09677419 0.90322581) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V12>=0.1694 86 32 M (0.62790698 0.37209302)  
#>      4) V36< 0.4927 72 19 M (0.73611111 0.26388889)  
#>        8) V47>=0.0365 65 12 M (0.81538462 0.18461538)  
#>         16) V54>=0.00455 57  5 M (0.91228070 0.08771930) *
#>         17) V54< 0.00455 8  1 R (0.12500000 0.87500000) *
#>        9) V47< 0.0365 7  0 R (0.00000000 1.00000000) *
#>      5) V36>=0.4927 14  1 R (0.07142857 0.92857143) *
#>    3) V12< 0.1694 53  9 R (0.16981132 0.83018868)  
#>      6) V26< 0.54215 8  2 M (0.75000000 0.25000000) *
#>      7) V26>=0.54215 45  3 R (0.06666667 0.93333333) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V45>=0.17155 51  5 M (0.90196078 0.09803922)  
#>      4) V33< 0.74045 43  0 M (1.00000000 0.00000000) *
#>      5) V33>=0.74045 8  3 R (0.37500000 0.62500000) *
#>    3) V45< 0.17155 88 22 R (0.25000000 0.75000000)  
#>      6) V49>=0.0519 14  3 M (0.78571429 0.21428571) *
#>      7) V49< 0.0519 74 11 R (0.14864865 0.85135135)  
#>       14) V48< 0.0227 12  5 M (0.58333333 0.41666667) *
#>       15) V48>=0.0227 62  4 R (0.06451613 0.93548387) *
#> 
#> 
#> $weights
#>  [1] 0.5834658 0.5847488 0.7788659 0.5853847 0.6830221 0.8539254 0.9424595
#>  [8] 0.5870975 0.4913670 0.4867231
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.0784645 5.4985952
#>   [2,] 1.5628389 5.0142208
#>   [3,] 2.5339958 4.0430639
#>   [4,] 1.6254815 4.9515782
#>   [5,] 2.1113100 4.4657497
#>   [6,] 2.4361522 4.1409075
#>   [7,] 2.0139314 4.5631283
#>   [8,] 0.0000000 6.5770597
#>   [9,] 1.7553120 4.8217477
#>  [10,] 2.7089071 3.8681526
#>  [11,] 2.2403224 4.3367373
#>  [12,] 2.4407961 4.1362636
#>  [13,] 0.5853847 5.9916750
#>  [14,] 1.7544939 4.8225658
#>  [15,] 2.1482236 4.4288361
#>  [16,] 0.6830221 5.8940377
#>  [17,] 2.0244886 4.5525711
#>  [18,] 2.1106741 4.4663857
#>  [19,] 2.6343702 3.9426896
#>  [20,] 0.8539254 5.7231343
#>  [21,] 1.6572863 4.9197734
#>  [22,] 1.3642506 5.2128091
#>  [23,] 0.5853847 5.9916750
#>  [24,] 1.9494291 4.6276306
#>  [25,] 2.5752508 4.0018089
#>  [26,] 0.0000000 6.5770597
#>  [27,] 0.0000000 6.5770597
#>  [28,] 1.5369475 5.0401123
#>  [29,] 1.4386742 5.1383855
#>  [30,] 0.4867231 6.0903366
#>  [31,] 1.1701335 5.4069262
#>  [32,] 1.7553120 4.8217477
#>  [33,] 2.2102304 4.3668294
#>  [34,] 1.3636147 5.2134450
#>  [35,] 1.0784645 5.4985952
#>  [36,] 0.6830221 5.8940377
#>  [37,] 0.8539254 5.7231343
#>  [38,] 1.2664878 5.3105719
#>  [39,] 1.6254815 4.9515782
#>  [40,] 0.0000000 6.5770597
#>  [41,] 0.4867231 6.0903366
#>  [42,] 0.4867231 6.0903366
#>  [43,] 1.6254815 4.9515782
#>  [44,] 1.4338265 5.1432332
#>  [45,] 0.4867231 6.0903366
#>  [46,] 1.3636147 5.2134450
#>  [47,] 0.6830221 5.8940377
#>  [48,] 1.3659634 5.2110964
#>  [49,] 0.0000000 6.5770597
#>  [50,] 0.8539254 5.7231343
#>  [51,] 1.1682146 5.4088451
#>  [52,] 1.4410229 5.1360369
#>  [53,] 1.0784645 5.4985952
#>  [54,] 2.2445779 4.3324818
#>  [55,] 2.1168486 4.4602112
#>  [56,] 2.7003143 3.8767454
#>  [57,] 2.2403224 4.3367373
#>  [58,] 1.6555736 4.9214861
#>  [59,] 2.4361522 4.1409075
#>  [60,] 2.6084193 3.9686404
#>  [61,] 1.0714719 5.5055878
#>  [62,] 3.3592238 3.2178359
#>  [63,] 1.9532550 4.6238048
#>  [64,] 1.4393101 5.1377496
#>  [65,] 2.0172923 4.5597675
#>  [66,] 5.8940377 0.6830221
#>  [67,] 4.4644667 2.1125930
#>  [68,] 3.7836265 2.7934332
#>  [69,] 3.9730997 2.6039601
#>  [70,] 5.3114707 1.2655890
#>  [71,] 4.4547275 2.1223322
#>  [72,] 3.9713869 2.6056728
#>  [73,] 5.2147281 1.3623317
#>  [74,] 4.2347190 2.3423407
#>  [75,] 5.5049519 1.0721078
#>  [76,] 6.5770597 0.0000000
#>  [77,] 5.4082092 1.1688505
#>  [78,] 4.6257116 1.9513481
#>  [79,] 5.5032392 1.0738205
#>  [80,] 4.3350246 2.2420351
#>  [81,] 5.1360369 1.4410229
#>  [82,] 6.5770597 0.0000000
#>  [83,] 5.3105719 1.2664878
#>  [84,] 6.5770597 0.0000000
#>  [85,] 5.2128091 1.3642506
#>  [86,] 4.3643673 2.2126924
#>  [87,] 5.2128091 1.3642506
#>  [88,] 4.2703496 2.3067101
#>  [89,] 4.8179218 1.7591379
#>  [90,] 5.8940377 0.6830221
#>  [91,] 5.1377496 1.4393101
#>  [92,] 5.1360369 1.4410229
#>  [93,] 5.8940377 0.6830221
#>  [94,] 5.2110964 1.3659634
#>  [95,] 4.7243733 1.8526864
#>  [96,] 4.6463826 1.9306772
#>  [97,] 5.7981938 0.7788659
#>  [98,] 4.6263475 1.9507122
#>  [99,] 5.5049519 1.0721078
#> [100,] 4.5506521 2.0264076
#> [101,] 4.5530008 2.0240589
#> [102,] 4.8198287 1.7572310
#> [103,] 4.0620222 2.5150375
#> [104,] 4.0620222 2.5150375
#> [105,] 5.2317673 1.3452924
#> [106,] 4.5487453 2.0283145
#> [107,] 5.5003080 1.0767518
#> [108,] 5.9916750 0.5853847
#> [109,] 5.0142208 1.5628389
#> [110,] 3.6467227 2.9303370
#> [111,] 5.4026707 1.1743891
#> [112,] 5.4026707 1.1743891
#> [113,] 4.8192049 1.7578548
#> [114,] 4.8201037 1.7569560
#> [115,] 4.6493138 1.9277459
#> [116,] 5.3105719 1.2664878
#> [117,] 4.5506521 2.0264076
#> [118,] 3.9678223 2.6092374
#> [119,] 3.7453843 2.8316754
#> [120,] 5.9935940 0.5834658
#> [121,] 6.5770597 0.0000000
#> [122,] 6.5770597 0.0000000
#> [123,] 6.5770597 0.0000000
#> [124,] 6.5770597 0.0000000
#> [125,] 5.4985952 1.0784645
#> [126,] 4.9197734 1.6572863
#> [127,] 5.0118722 1.5651876
#> [128,] 4.9221221 1.6549377
#> [129,] 4.9197734 1.6572863
#> [130,] 5.4985952 1.0784645
#> [131,] 6.0856927 0.4913670
#> [132,] 5.9899623 0.5870975
#> [133,] 4.8238488 1.7532109
#> [134,] 6.0856927 0.4913670
#> [135,] 4.8179218 1.7591379
#> [136,] 6.5770597 0.0000000
#> [137,] 5.8940377 0.6830221
#> [138,] 4.0645650 2.5124948
#> [139,] 4.7234744 1.8535853
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.16397365 0.83602635
#>   [2,] 0.23761969 0.76238031
#>   [3,] 0.38527790 0.61472210
#>   [4,] 0.24714411 0.75285589
#>   [5,] 0.32101122 0.67898878
#>   [6,] 0.37040141 0.62959859
#>   [7,] 0.30620543 0.69379457
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.26688401 0.73311599
#>  [10,] 0.41187206 0.58812794
#>  [11,] 0.34062674 0.65937326
#>  [12,] 0.37110749 0.62889251
#>  [13,] 0.08900402 0.91099598
#>  [14,] 0.26675962 0.73324038
#>  [15,] 0.32662371 0.67337629
#>  [16,] 0.10384915 0.89615085
#>  [17,] 0.30781059 0.69218941
#>  [18,] 0.32091454 0.67908546
#>  [19,] 0.40053919 0.59946081
#>  [20,] 0.12983391 0.87016609
#>  [21,] 0.25197982 0.74802018
#>  [22,] 0.20742561 0.79257439
#>  [23,] 0.08900402 0.91099598
#>  [24,] 0.29639827 0.70360173
#>  [25,] 0.39155046 0.60844954
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.23368306 0.76631694
#>  [29,] 0.21874124 0.78125876
#>  [30,] 0.07400314 0.92599686
#>  [31,] 0.17791135 0.82208865
#>  [32,] 0.26688401 0.73311599
#>  [33,] 0.33605143 0.66394857
#>  [34,] 0.20732892 0.79267108
#>  [35,] 0.16397365 0.83602635
#>  [36,] 0.10384915 0.89615085
#>  [37,] 0.12983391 0.87016609
#>  [38,] 0.19256140 0.80743860
#>  [39,] 0.24714411 0.75285589
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.07400314 0.92599686
#>  [42,] 0.07400314 0.92599686
#>  [43,] 0.24714411 0.75285589
#>  [44,] 0.21800418 0.78199582
#>  [45,] 0.07400314 0.92599686
#>  [46,] 0.20732892 0.79267108
#>  [47,] 0.10384915 0.89615085
#>  [48,] 0.20768602 0.79231398
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.12983391 0.87016609
#>  [51,] 0.17761958 0.82238042
#>  [52,] 0.21909834 0.78090166
#>  [53,] 0.16397365 0.83602635
#>  [54,] 0.34127376 0.65872624
#>  [55,] 0.32185333 0.67814667
#>  [56,] 0.41056558 0.58943442
#>  [57,] 0.34062674 0.65937326
#>  [58,] 0.25171941 0.74828059
#>  [59,] 0.37040141 0.62959859
#>  [60,] 0.39659353 0.60340647
#>  [61,] 0.16291047 0.83708953
#>  [62,] 0.51074856 0.48925144
#>  [63,] 0.29697997 0.70302003
#>  [64,] 0.21883793 0.78116207
#>  [65,] 0.30671643 0.69328357
#>  [66,] 0.89615085 0.10384915
#>  [67,] 0.67879370 0.32120630
#>  [68,] 0.57527629 0.42472371
#>  [69,] 0.60408448 0.39591552
#>  [70,] 0.80757526 0.19242474
#>  [71,] 0.67731292 0.32268708
#>  [72,] 0.60382407 0.39617593
#>  [73,] 0.79286616 0.20713384
#>  [74,] 0.64386203 0.35613797
#>  [75,] 0.83699284 0.16300716
#>  [76,] 1.00000000 0.00000000
#>  [77,] 0.82228373 0.17771627
#>  [78,] 0.70330996 0.29669004
#>  [79,] 0.83673243 0.16326757
#>  [80,] 0.65911285 0.34088715
#>  [81,] 0.78090166 0.21909834
#>  [82,] 1.00000000 0.00000000
#>  [83,] 0.80743860 0.19256140
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.79257439 0.20742561
#>  [86,] 0.66357423 0.33642577
#>  [87,] 0.79257439 0.20742561
#>  [88,] 0.64927943 0.35072057
#>  [89,] 0.73253430 0.26746570
#>  [90,] 0.89615085 0.10384915
#>  [91,] 0.78116207 0.21883793
#>  [92,] 0.78090166 0.21909834
#>  [93,] 0.89615085 0.10384915
#>  [94,] 0.79231398 0.20768602
#>  [95,] 0.71831084 0.28168916
#>  [96,] 0.70645285 0.29354715
#>  [97,] 0.88157841 0.11842159
#>  [98,] 0.70340665 0.29659335
#>  [99,] 0.83699284 0.16300716
#> [100,] 0.69189764 0.30810236
#> [101,] 0.69225474 0.30774526
#> [102,] 0.73282423 0.26717577
#> [103,] 0.61760457 0.38239543
#> [104,] 0.61760457 0.38239543
#> [105,] 0.79545687 0.20454313
#> [106,] 0.69160772 0.30839228
#> [107,] 0.83628676 0.16371324
#> [108,] 0.91099598 0.08900402
#> [109,] 0.76238031 0.23761969
#> [110,] 0.55446093 0.44553907
#> [111,] 0.82144163 0.17855837
#> [112,] 0.82144163 0.17855837
#> [113,] 0.73272938 0.26727062
#> [114,] 0.73286604 0.26713396
#> [115,] 0.70689852 0.29310148
#> [116,] 0.80743860 0.19256140
#> [117,] 0.69189764 0.30810236
#> [118,] 0.60328208 0.39671792
#> [119,] 0.56946181 0.43053819
#> [120,] 0.91128775 0.08871225
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 0.83602635 0.16397365
#> [126,] 0.74802018 0.25197982
#> [127,] 0.76202321 0.23797679
#> [128,] 0.74837728 0.25162272
#> [129,] 0.74802018 0.25197982
#> [130,] 0.83602635 0.16397365
#> [131,] 0.92529078 0.07470922
#> [132,] 0.91073557 0.08926443
#> [133,] 0.73343546 0.26656454
#> [134,] 0.92529078 0.07470922
#> [135,] 0.73253430 0.26746570
#> [136,] 1.00000000 0.00000000
#> [137,] 0.89615085 0.10384915
#> [138,] 0.61799119 0.38200881
#> [139,] 0.71817417 0.28182583
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "M" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  4.0525977  0.6862876  9.4271944  1.2206800  0.0000000  2.5762546 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  2.0520450  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  3.7639269  0.0000000  2.5845536  0.9962810  0.0000000  2.3982436 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.8109438  0.0000000  1.9141545  5.8175468  0.0000000  0.8198120  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  4.8396673  5.8356721  0.0000000  0.9876606  1.1920002  4.3407880 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  1.5503752  0.0000000  1.7494508  7.7678094  0.0000000  1.3197569 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  5.0700157 10.1758983  0.0000000  1.4366029  2.2830822  3.0442609  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  2.0782707  1.6914013  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  2.0984757  3.4182901 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V49        V12        V45        V37        V31        V48        V36 
#> 10.1758983  9.4271944  7.7678094  5.8356721  5.8175468  5.0700157  4.8396673 
#>        V40        V10        V23         V9        V52        V25        V15 
#>  4.3407880  4.0525977  3.7639269  3.4182901  3.0442609  2.5845536  2.5762546 
#>        V28        V51         V8        V54        V16        V30        V44 
#>  2.3982436  2.2830822  2.0984757  2.0782707  2.0520450  1.9141545  1.7494508 
#>        V55        V42        V50        V47        V13         V4        V26 
#>  1.6914013  1.5503752  1.4366029  1.3197569  1.2206800  1.1920002  0.9962810 
#>        V39        V33        V29        V11         V1        V14        V17 
#>  0.9876606  0.8198120  0.8109438  0.6862876  0.0000000  0.0000000  0.0000000 
#>        V18        V19         V2        V20        V21        V22        V24 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V27         V3        V32        V34        V35        V38        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V43        V46         V5        V53        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6        V60         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913