Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V48>=0.07485 75 13 M (0.82666667 0.17333333)  
#>     4) V9>=0.10015 65  5 M (0.92307692 0.07692308)  
#>       8) V14< 0.44365 57  0 M (1.00000000 0.00000000) *
#>       9) V14>=0.44365 8  3 R (0.37500000 0.62500000) *
#>     5) V9< 0.10015 10  2 R (0.20000000 0.80000000) *
#>   3) V48< 0.07485 64 15 R (0.23437500 0.76562500)  
#>     6) V21>=0.8107 13  3 M (0.76923077 0.23076923) *
#>     7) V21< 0.8107 51  5 R (0.09803922 0.90196078) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.19825 90 22 M (0.75555556 0.24444444)  
#>      4) V43>=0.18465 62  5 M (0.91935484 0.08064516)  
#>        8) V14< 0.5146 55  1 M (0.98181818 0.01818182) *
#>        9) V14>=0.5146 7  3 R (0.42857143 0.57142857) *
#>      5) V43< 0.18465 28 11 R (0.39285714 0.60714286)  
#>       10) V11>=0.2808 14  4 M (0.71428571 0.28571429) *
#>       11) V11< 0.2808 14  1 R (0.07142857 0.92857143) *
#>    3) V11< 0.19825 49  7 R (0.14285714 0.85714286)  
#>      6) V45>=0.16285 9  2 M (0.77777778 0.22222222) *
#>      7) V45< 0.16285 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 64 M (0.53956835 0.46043165)  
#>   2) V11>=0.19795 89 22 M (0.75280899 0.24719101)  
#>     4) V17< 0.90505 79 12 M (0.84810127 0.15189873)  
#>       8) V32>=0.08755 71  4 M (0.94366197 0.05633803) *
#>       9) V32< 0.08755 8  0 R (0.00000000 1.00000000) *
#>     5) V17>=0.90505 10  0 R (0.00000000 1.00000000) *
#>   3) V11< 0.19795 50  8 R (0.16000000 0.84000000)  
#>     6) V45>=0.1977 12  5 M (0.58333333 0.41666667) *
#>     7) V45< 0.1977 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 M (0.51079137 0.48920863)  
#>   2) V9>=0.1002 100 32 M (0.68000000 0.32000000)  
#>     4) V10>=0.12655 87 20 M (0.77011494 0.22988506)  
#>       8) V32>=0.2148 69  8 M (0.88405797 0.11594203) *
#>       9) V32< 0.2148 18  6 R (0.33333333 0.66666667) *
#>     5) V10< 0.12655 13  1 R (0.07692308 0.92307692) *
#>   3) V9< 0.1002 39  3 R (0.07692308 0.92307692) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V31< 0.39045 49  8 M (0.83673469 0.16326531)  
#>      4) V38>=0.2808 34  0 M (1.00000000 0.00000000) *
#>      5) V38< 0.2808 15  7 R (0.46666667 0.53333333) *
#>    3) V31>=0.39045 90 27 R (0.30000000 0.70000000)  
#>      6) V28>=0.958 13  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.958 77 14 R (0.18181818 0.81818182)  
#>       14) V43>=0.3404 18  5 M (0.72222222 0.27777778) *
#>       15) V43< 0.3404 59  1 R (0.01694915 0.98305085) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V36< 0.4781 92 26 M (0.71739130 0.28260870)  
#>      4) V11>=0.11155 84 18 M (0.78571429 0.21428571)  
#>        8) V47>=0.06475 61  5 M (0.91803279 0.08196721) *
#>        9) V47< 0.06475 23 10 R (0.43478261 0.56521739)  
#>         18) V44< 0.08945 8  0 M (1.00000000 0.00000000) *
#>         19) V44>=0.08945 15  2 R (0.13333333 0.86666667) *
#>      5) V11< 0.11155 8  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.4781 47  5 R (0.10638298 0.89361702)  
#>      6) V55< 0.0028 7  3 M (0.57142857 0.42857143) *
#>      7) V55>=0.0028 40  1 R (0.02500000 0.97500000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V4>=0.05265 55 12 M (0.78181818 0.21818182)  
#>      4) V57< 0.01465 42  1 M (0.97619048 0.02380952) *
#>      5) V57>=0.01465 13  2 R (0.15384615 0.84615385) *
#>    3) V4< 0.05265 84 25 R (0.29761905 0.70238095)  
#>      6) V23>=0.76675 23  7 M (0.69565217 0.30434783)  
#>       12) V2>=0.01255 16  1 M (0.93750000 0.06250000) *
#>       13) V2< 0.01255 7  1 R (0.14285714 0.85714286) *
#>      7) V23< 0.76675 61  9 R (0.14754098 0.85245902)  
#>       14) V28>=0.885 8  2 M (0.75000000 0.25000000) *
#>       15) V28< 0.885 53  3 R (0.05660377 0.94339623) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V31< 0.4353 55 14 M (0.74545455 0.25454545)  
#>      4) V44>=0.19055 29  0 M (1.00000000 0.00000000) *
#>      5) V44< 0.19055 26 12 R (0.46153846 0.53846154)  
#>       10) V33>=0.26895 11  0 M (1.00000000 0.00000000) *
#>       11) V33< 0.26895 15  1 R (0.06666667 0.93333333) *
#>    3) V31>=0.4353 84 24 R (0.28571429 0.71428571)  
#>      6) V9>=0.1236 45 21 M (0.53333333 0.46666667)  
#>       12) V51>=0.01445 25  3 M (0.88000000 0.12000000) *
#>       13) V51< 0.01445 20  2 R (0.10000000 0.90000000) *
#>      7) V9< 0.1236 39  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V12>=0.12715 104 30 M (0.71153846 0.28846154)  
#>      4) V9>=0.10655 80 14 M (0.82500000 0.17500000)  
#>        8) V46>=0.06695 65  5 M (0.92307692 0.07692308) *
#>        9) V46< 0.06695 15  6 R (0.40000000 0.60000000) *
#>      5) V9< 0.10655 24  8 R (0.33333333 0.66666667)  
#>       10) V31< 0.3822 9  1 M (0.88888889 0.11111111) *
#>       11) V31>=0.3822 15  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.12715 35  2 R (0.05714286 0.94285714) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V9>=0.1094 93 36 M (0.61290323 0.38709677)  
#>      4) V43>=0.24825 37  1 M (0.97297297 0.02702703) *
#>      5) V43< 0.24825 56 21 R (0.37500000 0.62500000)  
#>       10) V36< 0.154 16  3 M (0.81250000 0.18750000) *
#>       11) V36>=0.154 40  8 R (0.20000000 0.80000000)  
#>         22) V29< 0.3899 9  3 M (0.66666667 0.33333333) *
#>         23) V29>=0.3899 31  2 R (0.06451613 0.93548387) *
#>    3) V9< 0.1094 46  5 R (0.10869565 0.89130435)  
#>      6) V60< 0.00425 7  2 M (0.71428571 0.28571429) *
#>      7) V60>=0.00425 39  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.8302199 0.8385809 0.6262145 0.6625019 0.7171588 0.7474867
#>  [8] 0.9778697 0.6385012 0.6404247
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.2039210 5.2584762
#>   [2,] 2.1819541 5.2804431
#>   [3,] 2.7353602 4.7270371
#>   [4,] 3.4888064 3.9735909
#>   [5,] 0.9778697 6.4845276
#>   [6,] 0.8302199 6.6321774
#>   [7,] 2.9050872 4.5573101
#>   [8,] 2.4549517 5.0074455
#>   [9,] 3.0263594 4.4360378
#>  [10,] 0.7474867 6.7149105
#>  [11,] 2.2540837 5.2083136
#>  [12,] 2.3313026 5.1310946
#>  [13,] 3.0567121 4.4056851
#>  [14,] 0.6404247 6.8219726
#>  [15,] 1.6688008 5.7935965
#>  [16,] 1.3859879 6.0764094
#>  [17,] 1.3575835 6.1048138
#>  [18,] 2.2264923 5.2359050
#>  [19,] 1.3879114 6.0744859
#>  [20,] 0.9778697 6.4845276
#>  [21,] 1.4099886 6.0524087
#>  [22,] 2.1390991 5.3232981
#>  [23,] 1.5860676 5.8763297
#>  [24,] 0.7474867 6.7149105
#>  [25,] 0.7474867 6.7149105
#>  [26,] 1.3796607 6.0827366
#>  [27,] 2.2445088 5.2178885
#>  [28,] 0.6404247 6.8219726
#>  [29,] 1.6688008 5.7935965
#>  [30,] 1.5860676 5.8763297
#>  [31,] 2.3313026 5.1310946
#>  [32,] 0.6625019 6.7998954
#>  [33,] 1.6688008 5.7935965
#>  [34,] 0.6404247 6.8219726
#>  [35,] 0.6404247 6.8219726
#>  [36,] 0.7474867 6.7149105
#>  [37,] 2.7656486 4.6967487
#>  [38,] 2.1050702 5.3573271
#>  [39,] 0.0000000 7.4623973
#>  [40,] 0.6262145 6.8361828
#>  [41,] 1.3433733 6.1190240
#>  [42,] 0.6262145 6.8361828
#>  [43,] 0.6404247 6.8219726
#>  [44,] 0.0000000 7.4623973
#>  [45,] 0.0000000 7.4623973
#>  [46,] 1.4238638 6.0385335
#>  [47,] 2.1713505 5.2910467
#>  [48,] 1.4238638 6.0385335
#>  [49,] 2.3391788 5.1232184
#>  [50,] 2.8120943 4.6503030
#>  [51,] 2.6222991 4.8400982
#>  [52,] 1.7253564 5.7370409
#>  [53,] 2.1713505 5.2910467
#>  [54,] 3.3941572 4.0682401
#>  [55,] 1.2666391 6.1957582
#>  [56,] 2.1032965 5.3591007
#>  [57,] 1.9051403 5.5572570
#>  [58,] 2.1052200 5.3571773
#>  [59,] 1.2666391 6.1957582
#>  [60,] 0.0000000 7.4623973
#>  [61,] 0.7834391 6.6789581
#>  [62,] 5.9977517 1.4646455
#>  [63,] 6.8219726 0.6404247
#>  [64,] 5.4261942 2.0362030
#>  [65,] 5.3779551 2.0844422
#>  [66,] 5.3779551 2.0844422
#>  [67,] 5.3052569 2.1571403
#>  [68,] 6.0764094 1.3859879
#>  [69,] 5.4785993 1.9837979
#>  [70,] 5.2804431 2.1819541
#>  [71,] 5.5332563 1.9291410
#>  [72,] 6.8361828 0.6262145
#>  [73,] 6.7998954 0.6625019
#>  [74,] 3.6042295 3.8581678
#>  [75,] 6.4845276 0.9778697
#>  [76,] 7.4623973 0.0000000
#>  [77,] 5.4442354 2.0181618
#>  [78,] 5.3760316 2.0863657
#>  [79,] 4.5600219 2.9023753
#>  [80,] 5.3434611 2.1189362
#>  [81,] 6.7452385 0.7171588
#>  [82,] 4.7188259 2.7435714
#>  [83,] 5.7673688 1.6950285
#>  [84,] 6.1976816 1.2647156
#>  [85,] 5.4565222 2.0058751
#>  [86,] 7.4623973 0.0000000
#>  [87,] 6.8361828 0.6262145
#>  [88,] 6.7998954 0.6625019
#>  [89,] 6.7149105 0.7474867
#>  [90,] 4.7330361 2.7293611
#>  [91,] 5.1012515 2.3611457
#>  [92,] 5.1550953 2.3073019
#>  [93,] 6.8238961 0.6385012
#>  [94,] 6.4845276 0.9778697
#>  [95,] 5.3228133 2.1395839
#>  [96,] 6.1613942 1.3010030
#>  [97,] 4.8038108 2.6585865
#>  [98,] 6.7452385 0.7171588
#>  [99,] 6.4845276 0.9778697
#> [100,] 5.7673688 1.6950285
#> [101,] 6.7452385 0.7171588
#> [102,] 7.4623973 0.0000000
#> [103,] 6.0524087 1.4099886
#> [104,] 5.8220257 1.6403715
#> [105,] 6.0524087 1.4099886
#> [106,] 6.0524087 1.4099886
#> [107,] 4.6970837 2.7653136
#> [108,] 5.9617993 1.5005980
#> [109,] 5.3355849 2.1268124
#> [110,] 6.0527437 1.4096536
#> [111,] 6.1048138 1.3575835
#> [112,] 5.7010884 1.7613088
#> [113,] 7.4623973 0.0000000
#> [114,] 5.9314714 1.5309259
#> [115,] 5.9617993 1.5005980
#> [116,] 4.8157268 2.6466704
#> [117,] 7.4623973 0.0000000
#> [118,] 7.4623973 0.0000000
#> [119,] 7.4623973 0.0000000
#> [120,] 7.4623973 0.0000000
#> [121,] 7.4623973 0.0000000
#> [122,] 5.7370409 1.7253564
#> [123,] 7.4623973 0.0000000
#> [124,] 7.4623973 0.0000000
#> [125,] 6.6321774 0.8302199
#> [126,] 5.6543077 1.8080895
#> [127,] 6.6321774 0.8302199
#> [128,] 6.4845276 0.9778697
#> [129,] 6.4845276 0.9778697
#> [130,] 7.4623973 0.0000000
#> [131,] 6.0524087 1.4099886
#> [132,] 7.4623973 0.0000000
#> [133,] 6.6789581 0.7834391
#> [134,] 7.4623973 0.0000000
#> [135,] 6.1067373 1.3556600
#> [136,] 7.4623973 0.0000000
#> [137,] 5.2056017 2.2567955
#> [138,] 5.8220257 1.6403715
#> [139,] 6.0164563 1.4459410
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.29533687 0.70466313
#>   [2,] 0.29239319 0.70760681
#>   [3,] 0.36655247 0.63344753
#>   [4,] 0.46751818 0.53248182
#>   [5,] 0.13103962 0.86896038
#>   [6,] 0.11125377 0.88874623
#>   [7,] 0.38929677 0.61070323
#>   [8,] 0.32897629 0.67102371
#>   [9,] 0.40554789 0.59445211
#>  [10,] 0.10016710 0.89983290
#>  [11,] 0.30205892 0.69794108
#>  [12,] 0.31240666 0.68759334
#>  [13,] 0.40961531 0.59038469
#>  [14,] 0.08582023 0.91417977
#>  [15,] 0.22362797 0.77637203
#>  [16,] 0.18572958 0.81427042
#>  [17,] 0.18192324 0.81807676
#>  [18,] 0.29836153 0.70163847
#>  [19,] 0.18598734 0.81401266
#>  [20,] 0.13103962 0.86896038
#>  [21,] 0.18894579 0.81105421
#>  [22,] 0.28665039 0.71334961
#>  [23,] 0.21254130 0.78745870
#>  [24,] 0.10016710 0.89983290
#>  [25,] 0.10016710 0.89983290
#>  [26,] 0.18488170 0.81511830
#>  [27,] 0.30077584 0.69922416
#>  [28,] 0.08582023 0.91417977
#>  [29,] 0.22362797 0.77637203
#>  [30,] 0.21254130 0.78745870
#>  [31,] 0.31240666 0.68759334
#>  [32,] 0.08877869 0.91122131
#>  [33,] 0.22362797 0.77637203
#>  [34,] 0.08582023 0.91417977
#>  [35,] 0.08582023 0.91417977
#>  [36,] 0.10016710 0.89983290
#>  [37,] 0.37061127 0.62938873
#>  [38,] 0.28209034 0.71790966
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.08391599 0.91608401
#>  [41,] 0.18001899 0.81998101
#>  [42,] 0.08391599 0.91608401
#>  [43,] 0.08582023 0.91417977
#>  [44,] 0.00000000 1.00000000
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.19080515 0.80919485
#>  [47,] 0.29097225 0.70902775
#>  [48,] 0.19080515 0.80919485
#>  [49,] 0.31346212 0.68653788
#>  [50,] 0.37683524 0.62316476
#>  [51,] 0.35140170 0.64859830
#>  [52,] 0.23120672 0.76879328
#>  [53,] 0.29097225 0.70902775
#>  [54,] 0.45483469 0.54516531
#>  [55,] 0.16973622 0.83026378
#>  [56,] 0.28185266 0.71814734
#>  [57,] 0.25529870 0.74470130
#>  [58,] 0.28211042 0.71788958
#>  [59,] 0.16973622 0.83026378
#>  [60,] 0.00000000 1.00000000
#>  [61,] 0.10498492 0.89501508
#>  [62,] 0.80372989 0.19627011
#>  [63,] 0.91417977 0.08582023
#>  [64,] 0.72713822 0.27286178
#>  [65,] 0.72067392 0.27932608
#>  [66,] 0.72067392 0.27932608
#>  [67,] 0.71093199 0.28906801
#>  [68,] 0.81427042 0.18572958
#>  [69,] 0.73416077 0.26583923
#>  [70,] 0.70760681 0.29239319
#>  [71,] 0.74148509 0.25851491
#>  [72,] 0.91608401 0.08391599
#>  [73,] 0.91122131 0.08877869
#>  [74,] 0.48298547 0.51701453
#>  [75,] 0.86896038 0.13103962
#>  [76,] 1.00000000 0.00000000
#>  [77,] 0.72955583 0.27044417
#>  [78,] 0.72041616 0.27958384
#>  [79,] 0.61106663 0.38893337
#>  [80,] 0.71605154 0.28394846
#>  [81,] 0.90389700 0.09610300
#>  [82,] 0.63234719 0.36765281
#>  [83,] 0.77285738 0.22714262
#>  [84,] 0.83052154 0.16947846
#>  [85,] 0.73120231 0.26879769
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.91608401 0.08391599
#>  [88,] 0.91122131 0.08877869
#>  [89,] 0.89983290 0.10016710
#>  [90,] 0.63425143 0.36574857
#>  [91,] 0.68359421 0.31640579
#>  [92,] 0.69080955 0.30919045
#>  [93,] 0.91443753 0.08556247
#>  [94,] 0.86896038 0.13103962
#>  [95,] 0.71328464 0.28671536
#>  [96,] 0.82565883 0.17434117
#>  [97,] 0.64373560 0.35626440
#>  [98,] 0.90389700 0.09610300
#>  [99,] 0.86896038 0.13103962
#> [100,] 0.77285738 0.22714262
#> [101,] 0.90389700 0.09610300
#> [102,] 1.00000000 0.00000000
#> [103,] 0.81105421 0.18894579
#> [104,] 0.78018169 0.21981831
#> [105,] 0.81105421 0.18894579
#> [106,] 0.81105421 0.18894579
#> [107,] 0.62943362 0.37056638
#> [108,] 0.79891208 0.20108792
#> [109,] 0.71499609 0.28500391
#> [110,] 0.81109909 0.18890091
#> [111,] 0.81807676 0.18192324
#> [112,] 0.76397547 0.23602453
#> [113,] 1.00000000 0.00000000
#> [114,] 0.79484798 0.20515202
#> [115,] 0.79891208 0.20108792
#> [116,] 0.64533241 0.35466759
#> [117,] 1.00000000 0.00000000
#> [118,] 1.00000000 0.00000000
#> [119,] 1.00000000 0.00000000
#> [120,] 1.00000000 0.00000000
#> [121,] 1.00000000 0.00000000
#> [122,] 0.76879328 0.23120672
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 0.88874623 0.11125377
#> [126,] 0.75770661 0.24229339
#> [127,] 0.88874623 0.11125377
#> [128,] 0.86896038 0.13103962
#> [129,] 0.86896038 0.13103962
#> [130,] 1.00000000 0.00000000
#> [131,] 0.81105421 0.18894579
#> [132,] 1.00000000 0.00000000
#> [133,] 0.89501508 0.10498492
#> [134,] 1.00000000 0.00000000
#> [135,] 0.81833452 0.18166548
#> [136,] 1.00000000 0.00000000
#> [137,] 0.69757768 0.30242232
#> [138,] 0.78018169 0.21981831
#> [139,] 0.80623639 0.19376361
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  1.7800795 13.0544768  3.7452069  0.0000000  1.9483244  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  2.8004270  0.0000000  0.0000000  1.2022213  0.0000000  1.9123993 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  1.9618509  0.0000000  0.0000000  0.0000000  0.0000000  3.8866519 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.8471673  0.0000000  0.0000000  8.2447145  4.2265278  2.8275606  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  5.7926085  0.0000000  1.0259010  0.0000000  3.0466703  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  7.3677801  3.5030475  3.1711506  1.1136641  1.4631355 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  4.9642054  0.0000000  0.0000000  0.0000000  3.4575977  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.6672508  0.0000000  2.6247480  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  1.0142934  0.0000000  0.0000000 12.3503387 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 78 61 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11         V9        V31        V43        V36        V48        V32 
#> 13.0544768 12.3503387  8.2447145  7.3677801  5.7926085  4.9642054  4.2265278 
#>        V28        V12        V44        V51        V45         V4        V33 
#>  3.8866519  3.7452069  3.5030475  3.4575977  3.1711506  3.0466703  2.8275606 
#>        V17        V57        V23        V14        V21        V10        V47 
#>  2.8004270  2.6247480  1.9618509  1.9483244  1.9123993  1.7800795  1.4631355 
#>         V2        V46        V38        V60        V29        V55         V1 
#>  1.2022213  1.1136641  1.0259010  1.0142934  0.8471673  0.6672508  0.0000000 
#>        V13        V15        V16        V18        V19        V20        V22 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V25        V26        V27         V3        V30        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V37        V39        V40        V41        V42        V49 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V5        V50        V52        V53        V54        V56        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551