Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 67 M (0.51798561 0.48201439)  
#>   2) V11>=0.1998 76 15 M (0.80263158 0.19736842)  
#>     4) V39>=0.1016 69  9 M (0.86956522 0.13043478)  
#>       8) V6< 0.18085 61  4 M (0.93442623 0.06557377) *
#>       9) V6>=0.18085 8  3 R (0.37500000 0.62500000) *
#>     5) V39< 0.1016 7  1 R (0.14285714 0.85714286) *
#>   3) V11< 0.1998 63 11 R (0.17460317 0.82539683)  
#>     6) V45>=0.1691 15  6 M (0.60000000 0.40000000) *
#>     7) V45< 0.1691 48  2 R (0.04166667 0.95833333) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V51>=0.01365 83 24 M (0.71084337 0.28915663)  
#>      4) V11>=0.21635 53  5 M (0.90566038 0.09433962)  
#>        8) V12>=0.09485 46  1 M (0.97826087 0.02173913) *
#>        9) V12< 0.09485 7  3 R (0.42857143 0.57142857) *
#>      5) V11< 0.21635 30 11 R (0.36666667 0.63333333)  
#>       10) V4>=0.05565 11  2 M (0.81818182 0.18181818) *
#>       11) V4< 0.05565 19  2 R (0.10526316 0.89473684) *
#>    3) V51< 0.01365 56  7 R (0.12500000 0.87500000)  
#>      6) V31< 0.35405 9  3 M (0.66666667 0.33333333) *
#>      7) V31>=0.35405 47  1 R (0.02127660 0.97872340) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V48>=0.08475 64 12 M (0.81250000 0.18750000)  
#>      4) V16< 0.53875 56  5 M (0.91071429 0.08928571)  
#>        8) V36< 0.80795 47  0 M (1.00000000 0.00000000) *
#>        9) V36>=0.80795 9  4 R (0.44444444 0.55555556) *
#>      5) V16>=0.53875 8  1 R (0.12500000 0.87500000) *
#>    3) V48< 0.08475 75 23 R (0.30666667 0.69333333)  
#>      6) V22>=0.68675 33 12 M (0.63636364 0.36363636)  
#>       12) V10>=0.1318 24  3 M (0.87500000 0.12500000) *
#>       13) V10< 0.1318 9  0 R (0.00000000 1.00000000) *
#>      7) V22< 0.68675 42  2 R (0.04761905 0.95238095) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V5>=0.03825 106 40 M (0.62264151 0.37735849)  
#>      4) V49>=0.06085 33  1 M (0.96969697 0.03030303) *
#>      5) V49< 0.06085 73 34 R (0.46575342 0.53424658)  
#>       10) V29>=0.8496 16  0 M (1.00000000 0.00000000) *
#>       11) V29< 0.8496 57 18 R (0.31578947 0.68421053)  
#>         22) V18>=0.79245 13  1 M (0.92307692 0.07692308) *
#>         23) V18< 0.79245 44  6 R (0.13636364 0.86363636)  
#>           46) V59< 0.0025 7  3 M (0.57142857 0.42857143) *
#>           47) V59>=0.0025 37  2 R (0.05405405 0.94594595) *
#>    3) V5< 0.03825 33  2 R (0.06060606 0.93939394) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V9>=0.1795 68 18 M (0.73529412 0.26470588)  
#>      4) V20< 0.92755 61 11 M (0.81967213 0.18032787)  
#>        8) V36< 0.4781 39  1 M (0.97435897 0.02564103) *
#>        9) V36>=0.4781 22 10 M (0.54545455 0.45454545)  
#>         18) V17< 0.30705 12  0 M (1.00000000 0.00000000) *
#>         19) V17>=0.30705 10  0 R (0.00000000 1.00000000) *
#>      5) V20>=0.92755 7  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.1795 71 17 R (0.23943662 0.76056338)  
#>      6) V28>=0.98865 8  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.98865 63  9 R (0.14285714 0.85714286)  
#>       14) V21>=0.75445 15  7 M (0.53333333 0.46666667) *
#>       15) V21< 0.75445 48  1 R (0.02083333 0.97916667) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V48>=0.04905 93 31 M (0.66666667 0.33333333)  
#>      4) V10>=0.07795 81 20 M (0.75308642 0.24691358)  
#>        8) V17< 0.5512 55  4 M (0.92727273 0.07272727) *
#>        9) V17>=0.5512 26 10 R (0.38461538 0.61538462)  
#>         18) V19>=0.81 13  3 M (0.76923077 0.23076923) *
#>         19) V19< 0.81 13  0 R (0.00000000 1.00000000) *
#>      5) V10< 0.07795 12  1 R (0.08333333 0.91666667) *
#>    3) V48< 0.04905 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 R (0.4748201 0.5251799)  
#>   2) V52>=0.012 66 17 M (0.7424242 0.2575758)  
#>     4) V49>=0.0238 50  5 M (0.9000000 0.1000000) *
#>     5) V49< 0.0238 16  4 R (0.2500000 0.7500000) *
#>   3) V52< 0.012 73 17 R (0.2328767 0.7671233)  
#>     6) V45>=0.2867 11  2 M (0.8181818 0.1818182) *
#>     7) V45< 0.2867 62  8 R (0.1290323 0.8709677) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V36< 0.4154 83 25 M (0.69879518 0.30120482)  
#>      4) V53>=0.00815 51  4 M (0.92156863 0.07843137) *
#>      5) V53< 0.00815 32 11 R (0.34375000 0.65625000)  
#>       10) V26< 0.583 8  0 M (1.00000000 0.00000000) *
#>       11) V26>=0.583 24  3 R (0.12500000 0.87500000) *
#>    3) V36>=0.4154 56  9 R (0.16071429 0.83928571)  
#>      6) V1>=0.0482 8  1 M (0.87500000 0.12500000) *
#>      7) V1< 0.0482 48  2 R (0.04166667 0.95833333) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V28>=0.911 33  0 M (1.00000000 0.00000000) *
#>    3) V28< 0.911 106 40 R (0.37735849 0.62264151)  
#>      6) V51>=0.0129 61 25 M (0.59016393 0.40983607)  
#>       12) V12>=0.1466 43  8 M (0.81395349 0.18604651)  
#>         24) V1>=0.02035 34  0 M (1.00000000 0.00000000) *
#>         25) V1< 0.02035 9  1 R (0.11111111 0.88888889) *
#>       13) V12< 0.1466 18  1 R (0.05555556 0.94444444) *
#>      7) V51< 0.0129 45  4 R (0.08888889 0.91111111)  
#>       14) V1< 0.0118 7  3 M (0.57142857 0.42857143) *
#>       15) V1>=0.0118 38  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.5251799 0.4748201)  
#>    2) V21>=0.55655 83 24 M (0.7108434 0.2891566)  
#>      4) V51>=0.012 50  5 M (0.9000000 0.1000000) *
#>      5) V51< 0.012 33 14 R (0.4242424 0.5757576)  
#>       10) V52>=0.0098 14  2 M (0.8571429 0.1428571) *
#>       11) V52< 0.0098 19  2 R (0.1052632 0.8947368) *
#>    3) V21< 0.55655 56 14 R (0.2500000 0.7500000)  
#>      6) V36< 0.1684 7  0 M (1.0000000 0.0000000) *
#>      7) V36>=0.1684 49  7 R (0.1428571 0.8571429)  
#>       14) V48>=0.12695 12  5 M (0.5833333 0.4166667) *
#>       15) V48< 0.12695 37  0 R (0.0000000 1.0000000) *
#> 
#> 
#> $weights
#>  [1] 0.8090480 0.7469430 0.8088468 0.7122882 0.8699710 0.7435987 0.6949277
#>  [8] 0.5350824 0.6918781 0.7201147
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.2727614 5.0599371
#>   [2,] 1.5822591 5.7504394
#>   [3,] 2.8687668 4.4639317
#>   [4,] 2.9055841 4.4271144
#>   [5,] 2.9610443 4.3716542
#>   [6,] 0.5350824 6.7976161
#>   [7,] 2.1251681 5.2075304
#>   [8,] 2.8687668 4.4639317
#>   [9,] 1.5526466 5.7800519
#>  [10,] 0.7469430 6.5857555
#>  [11,] 1.4388211 5.8938774
#>  [12,] 2.9914677 4.3412308
#>  [13,] 1.4418707 5.8908278
#>  [14,] 0.6918781 6.6408204
#>  [15,] 2.2759045 5.0567940
#>  [16,] 2.2238892 5.1088093
#>  [17,] 2.7309362 4.6017623
#>  [18,] 2.3097729 5.0229256
#>  [19,] 0.7435987 6.5890998
#>  [20,] 0.8090480 6.5236505
#>  [21,] 1.4592312 5.8734673
#>  [22,] 2.3097729 5.0229256
#>  [23,] 2.0908722 5.2418263
#>  [24,] 1.9739035 5.3587950
#>  [25,] 0.8699710 6.4627275
#>  [26,] 0.8088468 6.5238517
#>  [27,] 1.9218882 5.4108103
#>  [28,] 0.7435987 6.5890998
#>  [29,] 0.8699710 6.4627275
#>  [30,] 0.6918781 6.6408204
#>  [31,] 0.8090480 6.5236505
#>  [32,] 1.5009261 5.8317724
#>  [33,] 3.7452254 3.5874731
#>  [34,] 2.4257607 4.9069378
#>  [35,] 0.0000000 7.3326985
#>  [36,] 0.7435987 6.5890998
#>  [37,] 0.0000000 7.3326985
#>  [38,] 1.5007249 5.8319736
#>  [39,] 0.7435987 6.5890998
#>  [40,] 2.1477649 5.1849336
#>  [41,] 2.2443235 5.0883750
#>  [42,] 0.0000000 7.3326985
#>  [43,] 1.5211350 5.8115635
#>  [44,] 1.4041663 5.9285322
#>  [45,] 0.7435987 6.5890998
#>  [46,] 0.0000000 7.3326985
#>  [47,] 0.0000000 7.3326985
#>  [48,] 0.0000000 7.3326985
#>  [49,] 2.2162639 5.1164346
#>  [50,] 0.6949277 6.6377708
#>  [51,] 2.1511093 5.1815892
#>  [52,] 2.2995896 5.0331089
#>  [53,] 3.1112207 4.2214778
#>  [54,] 2.8200958 4.5126027
#>  [55,] 2.1251681 5.2075304
#>  [56,] 1.5900857 5.7426128
#>  [57,] 1.6135696 5.7191289
#>  [58,] 2.8609402 4.4717583
#>  [59,] 1.4324029 5.9002956
#>  [60,] 1.4670577 5.8656408
#>  [61,] 1.4324029 5.9002956
#>  [62,] 1.5291627 5.8035358
#>  [63,] 0.7122882 6.6204103
#>  [64,] 0.0000000 7.3326985
#>  [65,] 1.9470753 5.3856232
#>  [66,] 1.5291627 5.8035358
#>  [67,] 0.7122882 6.6204103
#>  [68,] 5.7677998 1.5648987
#>  [69,] 5.0853253 2.2473731
#>  [70,] 5.6538808 1.6788177
#>  [71,] 5.2022940 2.1304045
#>  [72,] 5.0883750 2.2443235
#>  [73,] 5.8317724 1.5009261
#>  [74,] 5.0679649 2.2647336
#>  [75,] 5.0881738 2.2445247
#>  [76,] 6.1026884 1.2300101
#>  [77,] 6.6408204 0.6918781
#>  [78,] 5.8421569 1.4905416
#>  [79,] 7.3326985 0.0000000
#>  [80,] 5.1370459 2.1956526
#>  [81,] 6.6377708 0.6949277
#>  [82,] 6.5236505 0.8090480
#>  [83,] 5.1196854 2.2130131
#>  [84,] 7.3326985 0.0000000
#>  [85,] 6.4627275 0.8699710
#>  [86,] 4.5499482 2.7827503
#>  [87,] 5.8115635 1.5211350
#>  [88,] 6.0853279 1.2473706
#>  [89,] 5.0585613 2.2741372
#>  [90,] 5.6538808 1.6788177
#>  [91,] 6.6204103 0.7122882
#>  [92,] 6.5236505 0.8090480
#>  [93,] 5.7708495 1.5618490
#>  [94,] 5.7426128 1.5900857
#>  [95,] 6.4627275 0.8699710
#>  [96,] 6.6408204 0.6918781
#>  [97,] 6.6204103 0.7122882
#>  [98,] 6.6377708 0.6949277
#>  [99,] 5.0851241 2.2475744
#> [100,] 6.5890998 0.7435987
#> [101,] 4.4617692 2.8709293
#> [102,] 5.1989497 2.1337488
#> [103,] 6.0506731 1.2820254
#> [104,] 6.1057380 1.2269605
#> [105,] 6.0506731 1.2820254
#> [106,] 5.9276451 1.4050534
#> [107,] 6.7976161 0.5350824
#> [108,] 5.9276451 1.4050534
#> [109,] 4.6356307 2.6970678
#> [110,] 6.1026884 1.2300101
#> [111,] 5.3383849 1.9943136
#> [112,] 5.1840465 2.1486520
#> [113,] 4.6389750 2.6937235
#> [114,] 5.3934498 1.9392487
#> [115,] 5.7769087 1.5557898
#> [116,] 5.9285322 1.4041663
#> [117,] 4.4614745 2.8712240
#> [118,] 5.9885681 1.3441304
#> [119,] 6.7976161 0.5350824
#> [120,] 6.7976161 0.5350824
#> [121,] 6.7976161 0.5350824
#> [122,] 6.5236505 0.8090480
#> [123,] 6.7976161 0.5350824
#> [124,] 5.8938774 1.4388211
#> [125,] 6.1026884 1.2300101
#> [126,] 7.3326985 0.0000000
#> [127,] 6.7976161 0.5350824
#> [128,] 5.3557454 1.9769531
#> [129,] 6.4627275 0.8699710
#> [130,] 4.5468986 2.7857999
#> [131,] 6.4627275 0.8699710
#> [132,] 6.0853279 1.2473706
#> [133,] 5.3904002 1.9422983
#> [134,] 6.6204103 0.7122882
#> [135,] 6.7976161 0.5350824
#> [136,] 6.1026884 1.2300101
#> [137,] 6.6377708 0.6949277
#> [138,] 5.8908278 1.4418707
#> [139,] 4.3728360 2.9598625
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.30994884 0.69005116
#>   [2,] 0.21578129 0.78421871
#>   [3,] 0.39122934 0.60877066
#>   [4,] 0.39625031 0.60374969
#>   [5,] 0.40381373 0.59618627
#>   [6,] 0.07297210 0.92702790
#>   [7,] 0.28982074 0.71017926
#>   [8,] 0.39122934 0.60877066
#>   [9,] 0.21174287 0.78825713
#>  [10,] 0.10186468 0.89813532
#>  [11,] 0.19621986 0.80378014
#>  [12,] 0.40796273 0.59203727
#>  [13,] 0.19663575 0.80336425
#>  [14,] 0.09435518 0.90564482
#>  [15,] 0.31037748 0.68962252
#>  [16,] 0.30328388 0.69671612
#>  [17,] 0.37243263 0.62756737
#>  [18,] 0.31499630 0.68500370
#>  [19,] 0.10140860 0.89859140
#>  [20,] 0.11033428 0.88966572
#>  [21,] 0.19900329 0.80099671
#>  [22,] 0.31499630 0.68500370
#>  [23,] 0.28514362 0.71485638
#>  [24,] 0.26919196 0.73080804
#>  [25,] 0.11864267 0.88135733
#>  [26,] 0.11030684 0.88969316
#>  [27,] 0.26209836 0.73790164
#>  [28,] 0.10140860 0.89859140
#>  [29,] 0.11864267 0.88135733
#>  [30,] 0.09435518 0.90564482
#>  [31,] 0.11033428 0.88966572
#>  [32,] 0.20468946 0.79531054
#>  [33,] 0.51075677 0.48924323
#>  [34,] 0.33081419 0.66918581
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.10140860 0.89859140
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.20466202 0.79533798
#>  [39,] 0.10140860 0.89859140
#>  [40,] 0.29290239 0.70709761
#>  [41,] 0.30607061 0.69392939
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.20744545 0.79255455
#>  [44,] 0.19149380 0.80850620
#>  [45,] 0.10140860 0.89859140
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.30224397 0.69775603
#>  [50,] 0.09477107 0.90522893
#>  [51,] 0.29335848 0.70664152
#>  [52,] 0.31360755 0.68639245
#>  [53,] 0.42429409 0.57570591
#>  [54,] 0.38459181 0.61540819
#>  [55,] 0.28982074 0.71017926
#>  [56,] 0.21684864 0.78315136
#>  [57,] 0.22005127 0.77994873
#>  [58,] 0.39016199 0.60983801
#>  [59,] 0.19534458 0.80465542
#>  [60,] 0.20007064 0.79992936
#>  [61,] 0.19534458 0.80465542
#>  [62,] 0.20854024 0.79145976
#>  [63,] 0.09713862 0.90286138
#>  [64,] 0.00000000 1.00000000
#>  [65,] 0.26553325 0.73446675
#>  [66,] 0.20854024 0.79145976
#>  [67,] 0.09713862 0.90286138
#>  [68,] 0.78658625 0.21341375
#>  [69,] 0.69351349 0.30648651
#>  [70,] 0.77105049 0.22894951
#>  [71,] 0.70946515 0.29053485
#>  [72,] 0.69392939 0.30607061
#>  [73,] 0.79531054 0.20468946
#>  [74,] 0.69114595 0.30885405
#>  [75,] 0.69390195 0.30609805
#>  [76,] 0.83225683 0.16774317
#>  [77,] 0.90564482 0.09435518
#>  [78,] 0.79672673 0.20327327
#>  [79,] 1.00000000 0.00000000
#>  [80,] 0.70056691 0.29943309
#>  [81,] 0.90522893 0.09477107
#>  [82,] 0.88966572 0.11033428
#>  [83,] 0.69819936 0.30180064
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.88135733 0.11864267
#>  [86,] 0.62050120 0.37949880
#>  [87,] 0.79255455 0.20744545
#>  [88,] 0.82988928 0.17011072
#>  [89,] 0.68986353 0.31013647
#>  [90,] 0.77105049 0.22894951
#>  [91,] 0.90286138 0.09713862
#>  [92,] 0.88966572 0.11033428
#>  [93,] 0.78700215 0.21299785
#>  [94,] 0.78315136 0.21684864
#>  [95,] 0.88135733 0.11864267
#>  [96,] 0.90564482 0.09435518
#>  [97,] 0.90286138 0.09713862
#>  [98,] 0.90522893 0.09477107
#>  [99,] 0.69348605 0.30651395
#> [100,] 0.89859140 0.10140860
#> [101,] 0.60847575 0.39152425
#> [102,] 0.70900907 0.29099093
#> [103,] 0.82516322 0.17483678
#> [104,] 0.83267272 0.16732728
#> [105,] 0.82516322 0.17483678
#> [106,] 0.80838523 0.19161477
#> [107,] 0.92702790 0.07297210
#> [108,] 0.80838523 0.19161477
#> [109,] 0.63218618 0.36781382
#> [110,] 0.83225683 0.16774317
#> [111,] 0.72802460 0.27197540
#> [112,] 0.70697663 0.29302337
#> [113,] 0.63264226 0.36735774
#> [114,] 0.73553410 0.26446590
#> [115,] 0.78782848 0.21217152
#> [116,] 0.80850620 0.19149380
#> [117,] 0.60843556 0.39156444
#> [118,] 0.81669362 0.18330638
#> [119,] 0.92702790 0.07297210
#> [120,] 0.92702790 0.07297210
#> [121,] 0.92702790 0.07297210
#> [122,] 0.88966572 0.11033428
#> [123,] 0.92702790 0.07297210
#> [124,] 0.80378014 0.19621986
#> [125,] 0.83225683 0.16774317
#> [126,] 1.00000000 0.00000000
#> [127,] 0.92702790 0.07297210
#> [128,] 0.73039215 0.26960785
#> [129,] 0.88135733 0.11864267
#> [130,] 0.62008531 0.37991469
#> [131,] 0.88135733 0.11864267
#> [132,] 0.82988928 0.17011072
#> [133,] 0.73511821 0.26488179
#> [134,] 0.90286138 0.09713862
#> [135,] 0.92702790 0.07297210
#> [136,] 0.83225683 0.16774317
#> [137,] 0.90522893 0.09477107
#> [138,] 0.80336425 0.19663575
#> [139,] 0.59634744 0.40365256
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 4.1420260 4.0173031 8.0716861 3.4210342 0.0000000 0.0000000 0.0000000 1.8624253 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 4.5882220 2.3571931 1.5238791 0.0000000 1.9556418 4.1163217 2.7606246 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 1.3097045 0.0000000 6.0141438 2.2196908 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 1.2521882 0.0000000 0.0000000 0.0000000 0.0000000 6.6910705 0.0000000 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 1.4468152 1.4092165 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 3.1787224 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 0.0000000 9.4021357 4.0867210 3.0169869 0.0000000 8.6929325 5.0807595 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 1.8716778 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.5979958 0.9541484 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 3.9587333 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V48       V51       V11       V36       V28       V52       V17        V1 
#> 9.4021357 8.6929325 8.0716861 6.6910705 6.0141438 5.0807595 4.5882220 4.1420260 
#>       V21       V49       V10        V9       V12       V45        V5       V22 
#> 4.1163217 4.0867210 4.0173031 3.9587333 3.4210342 3.1787224 3.0169869 2.7606246 
#>       V18       V29       V20       V53       V16       V19       V39        V4 
#> 2.3571931 2.2196908 1.9556418 1.8716778 1.8624253 1.5238791 1.4468152 1.4092165 
#>       V26       V31        V6       V59       V13       V14       V15        V2 
#> 1.3097045 1.2521882 0.9541484 0.5979958 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V23       V24       V25       V27        V3       V30       V32       V33 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V34       V35       V37       V38       V40       V41       V42       V43 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V44       V46       V47       V50       V54       V55       V56       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V58       V60        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986