Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V36< 0.48825 85 28 M (0.67058824 0.32941176)  
#>      4) V48>=0.04695 61  9 M (0.85245902 0.14754098)  
#>        8) V11>=0.11855 54  4 M (0.92592593 0.07407407) *
#>        9) V11< 0.11855 7  2 R (0.28571429 0.71428571) *
#>      5) V48< 0.04695 24  5 R (0.20833333 0.79166667)  
#>       10) V11>=0.28855 9  4 M (0.55555556 0.44444444) *
#>       11) V11< 0.28855 15  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.48825 54  9 R (0.16666667 0.83333333)  
#>      6) V45>=0.3571 9  1 M (0.88888889 0.11111111) *
#>      7) V45< 0.3571 45  1 R (0.02222222 0.97777778) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V10>=0.18455 70 16 M (0.77142857 0.22857143)  
#>      4) V36< 0.38435 43  2 M (0.95348837 0.04651163) *
#>      5) V36>=0.38435 27 13 R (0.48148148 0.51851852)  
#>       10) V17< 0.205 13  1 M (0.92307692 0.07692308) *
#>       11) V17>=0.205 14  1 R (0.07142857 0.92857143) *
#>    3) V10< 0.18455 69 15 R (0.21739130 0.78260870)  
#>      6) V21>=0.8058 15  3 M (0.80000000 0.20000000) *
#>      7) V21< 0.8058 54  3 R (0.05555556 0.94444444) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 64 R (0.46043165 0.53956835)  
#>   2) V10>=0.1506 80 26 M (0.67500000 0.32500000)  
#>     4) V7< 0.2137 65 12 M (0.81538462 0.18461538)  
#>       8) V21>=0.37595 53  4 M (0.92452830 0.07547170) *
#>       9) V21< 0.37595 12  4 R (0.33333333 0.66666667) *
#>     5) V7>=0.2137 15  1 R (0.06666667 0.93333333) *
#>   3) V10< 0.1506 59 10 R (0.16949153 0.83050847)  
#>     6) V54>=0.0158 10  1 M (0.90000000 0.10000000) *
#>     7) V54< 0.0158 49  1 R (0.02040816 0.97959184) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 R (0.49640288 0.50359712)  
#>   2) V11>=0.2049 78 22 M (0.71794872 0.28205128)  
#>     4) V46>=0.06915 60  7 M (0.88333333 0.11666667)  
#>       8) V15< 0.5777 52  2 M (0.96153846 0.03846154) *
#>       9) V15>=0.5777 8  3 R (0.37500000 0.62500000) *
#>     5) V46< 0.06915 18  3 R (0.16666667 0.83333333) *
#>   3) V11< 0.2049 61 13 R (0.21311475 0.78688525)  
#>     6) V20>=0.8427 10  2 M (0.80000000 0.20000000) *
#>     7) V20< 0.8427 51  5 R (0.09803922 0.90196078) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V11>=0.1089 113 40 M (0.64601770 0.35398230)  
#>      4) V48>=0.076 54  7 M (0.87037037 0.12962963)  
#>        8) V50>=0.0186 45  2 M (0.95555556 0.04444444) *
#>        9) V50< 0.0186 9  4 R (0.44444444 0.55555556) *
#>      5) V48< 0.076 59 26 R (0.44067797 0.55932203)  
#>       10) V22>=0.68675 31  7 M (0.77419355 0.22580645)  
#>         20) V1>=0.01485 22  1 M (0.95454545 0.04545455) *
#>         21) V1< 0.01485 9  3 R (0.33333333 0.66666667) *
#>       11) V22< 0.68675 28  2 R (0.07142857 0.92857143) *
#>    3) V11< 0.1089 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V48>=0.04695 94 32 M (0.65957447 0.34042553)  
#>      4) V37< 0.35815 53  5 M (0.90566038 0.09433962)  
#>        8) V3>=0.01215 46  0 M (1.00000000 0.00000000) *
#>        9) V3< 0.01215 7  2 R (0.28571429 0.71428571) *
#>      5) V37>=0.35815 41 14 R (0.34146341 0.65853659)  
#>       10) V29>=0.73865 8  0 M (1.00000000 0.00000000) *
#>       11) V29< 0.73865 33  6 R (0.18181818 0.81818182)  
#>         22) V18< 0.26275 7  2 M (0.71428571 0.28571429) *
#>         23) V18>=0.26275 26  1 R (0.03846154 0.96153846) *
#>    3) V48< 0.04695 45  7 R (0.15555556 0.84444444)  
#>      6) V22>=0.96205 7  1 M (0.85714286 0.14285714) *
#>      7) V22< 0.96205 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V11>=0.1117 101 35 M (0.65346535 0.34653465)  
#>      4) V4>=0.0474 59  9 M (0.84745763 0.15254237) *
#>      5) V4< 0.0474 42 16 R (0.38095238 0.61904762)  
#>       10) V22>=0.6993 18  4 M (0.77777778 0.22222222) *
#>       11) V22< 0.6993 24  2 R (0.08333333 0.91666667) *
#>    3) V11< 0.1117 38  2 R (0.05263158 0.94736842) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V31< 0.4569 74 23 M (0.68918919 0.31081081)  
#>      4) V42>=0.0851 64 13 M (0.79687500 0.20312500)  
#>        8) V37< 0.5297 42  1 M (0.97619048 0.02380952) *
#>        9) V37>=0.5297 22 10 R (0.45454545 0.54545455)  
#>         18) V24< 0.50945 10  0 M (1.00000000 0.00000000) *
#>         19) V24>=0.50945 12  0 R (0.00000000 1.00000000) *
#>      5) V42< 0.0851 10  0 R (0.00000000 1.00000000) *
#>    3) V31>=0.4569 65 15 R (0.23076923 0.76923077)  
#>      6) V49>=0.0654 13  3 M (0.76923077 0.23076923) *
#>      7) V49< 0.0654 52  5 R (0.09615385 0.90384615)  
#>       14) V28>=0.9585 8  3 M (0.62500000 0.37500000) *
#>       15) V28< 0.9585 44  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V52>=0.0094 77 28 M (0.63636364 0.36363636)  
#>      4) V11>=0.10775 66 17 M (0.74242424 0.25757576)  
#>        8) V15< 0.61845 58 10 M (0.82758621 0.17241379)  
#>         16) V9>=0.07485 46  3 M (0.93478261 0.06521739) *
#>         17) V9< 0.07485 12  5 R (0.41666667 0.58333333) *
#>        9) V15>=0.61845 8  1 R (0.12500000 0.87500000) *
#>      5) V11< 0.10775 11  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0094 62  9 R (0.14516129 0.85483871) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 67 M (0.51798561 0.48201439)  
#>   2) V12>=0.19455 94 30 M (0.68085106 0.31914894)  
#>     4) V32< 0.5372 76 14 M (0.81578947 0.18421053)  
#>       8) V35< 0.3663 59  4 M (0.93220339 0.06779661) *
#>       9) V35>=0.3663 17  7 R (0.41176471 0.58823529) *
#>     5) V32>=0.5372 18  2 R (0.11111111 0.88888889) *
#>   3) V12< 0.19455 45  8 R (0.17777778 0.82222222)  
#>     6) V55< 0.0041 13  5 M (0.61538462 0.38461538) *
#>     7) V55>=0.0041 32  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.8916956 0.7359500 0.7026978 0.5814024 0.7305775 0.6994132 0.6813128
#>  [8] 0.8349691 0.6447131 0.5421797
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.4332752 5.6116359
#>   [2,] 2.1151881 4.9297230
#>   [3,] 2.3089584 4.7359528
#>   [4,] 1.5343823 5.5105288
#>   [5,] 2.4293625 4.6155486
#>   [6,] 2.1698253 4.8750858
#>   [7,] 2.0287236 5.0161875
#>   [8,] 2.8550193 4.1898918
#>   [9,] 1.4021110 5.6428002
#>  [10,] 1.5376669 5.5072442
#>  [11,] 1.3807260 5.6641852
#>  [12,] 1.5376669 5.5072442
#>  [13,] 1.9621284 5.0827828
#>  [14,] 1.9255287 5.1193825
#>  [15,] 2.0566033 4.9883078
#>  [16,] 2.8511381 4.1937730
#>  [17,] 1.9566929 5.0882182
#>  [18,] 2.4572422 4.5876689
#>  [19,] 1.5343823 5.5105288
#>  [20,] 0.6994132 6.3454980
#>  [21,] 0.5421797 6.5027314
#>  [22,] 1.5343823 5.5105288
#>  [23,] 0.5421797 6.5027314
#>  [24,] 2.6959904 4.3489207
#>  [25,] 1.9229057 5.1220055
#>  [26,] 0.8349691 6.2099420
#>  [27,] 2.8550193 4.1898918
#>  [28,] 2.2102597 4.8346514
#>  [29,] 2.1544108 4.8905003
#>  [30,] 2.0287236 5.0161875
#>  [31,] 0.6813128 6.3635984
#>  [32,] 1.3806631 5.6642481
#>  [33,] 0.0000000 7.0449111
#>  [34,] 0.0000000 7.0449111
#>  [35,] 1.2808156 5.7640956
#>  [36,] 1.7266648 5.3182464
#>  [37,] 2.1145880 4.9303231
#>  [38,] 0.7305775 6.3143337
#>  [39,] 1.5162819 5.5286292
#>  [40,] 1.2627152 5.7821960
#>  [41,] 0.6994132 6.3454980
#>  [42,] 0.0000000 7.0449111
#>  [43,] 2.4260779 4.6188332
#>  [44,] 1.4338753 5.6110358
#>  [45,] 0.0000000 7.0449111
#>  [46,] 0.0000000 7.0449111
#>  [47,] 1.2234925 5.8214186
#>  [48,] 1.3840106 5.6609006
#>  [49,] 1.3771489 5.6677623
#>  [50,] 1.4021110 5.6428002
#>  [51,] 0.6994132 6.3454980
#>  [52,] 1.4021110 5.6428002
#>  [53,] 0.5421797 6.5027314
#>  [54,] 1.4796822 5.5652289
#>  [55,] 0.8349691 6.2099420
#>  [56,] 0.8349691 6.2099420
#>  [57,] 2.2688445 4.7760667
#>  [58,] 0.7359500 6.3089611
#>  [59,] 1.3173524 5.7275588
#>  [60,] 2.5589453 4.4859659
#>  [61,] 1.9775429 5.0673683
#>  [62,] 1.3173524 5.7275588
#>  [63,] 1.4353632 5.6095480
#>  [64,] 1.3840106 5.6609006
#>  [65,] 2.6567677 4.3881434
#>  [66,] 2.7890391 4.2558720
#>  [67,] 2.2189797 4.8259314
#>  [68,] 0.6813128 6.3635984
#>  [69,] 0.0000000 7.0449111
#>  [70,] 0.5421797 6.5027314
#>  [71,] 0.0000000 7.0449111
#>  [72,] 0.7026978 6.3422134
#>  [73,] 4.9068502 2.1380609
#>  [74,] 5.5783837 1.4665275
#>  [75,] 5.0882182 1.9566929
#>  [76,] 5.0882182 1.9566929
#>  [77,] 5.0863599 1.9585512
#>  [78,] 3.8412528 3.2036583
#>  [79,] 5.0335181 2.0113930
#>  [80,] 5.0302335 2.0146776
#>  [81,] 4.4559074 2.5890037
#>  [82,] 4.2669096 2.7780016
#>  [83,] 4.9271000 2.1178111
#>  [84,] 4.8058047 2.2391065
#>  [85,] 5.6110358 1.4338753
#>  [86,] 7.0449111 0.0000000
#>  [87,] 6.4635087 0.5814024
#>  [88,] 4.3834325 2.6614786
#>  [89,] 7.0449111 0.0000000
#>  [90,] 6.4001981 0.6447131
#>  [91,] 7.0449111 0.0000000
#>  [92,] 6.3454980 0.6994132
#>  [93,] 6.5027314 0.5421797
#>  [94,] 5.0882182 1.9566929
#>  [95,] 5.8187957 1.2261155
#>  [96,] 5.8187957 1.2261155
#>  [97,] 5.1220684 1.9228428
#>  [98,] 5.0673683 1.9775429
#>  [99,] 4.9083380 2.1365731
#> [100,] 4.9371848 2.1077263
#> [101,] 5.1853790 1.8595321
#> [102,] 5.0828457 1.9620655
#> [103,] 4.9969813 2.0479298
#> [104,] 6.5027314 0.5421797
#> [105,] 7.0449111 0.0000000
#> [106,] 5.2766160 1.7682952
#> [107,] 5.5718131 1.4730980
#> [108,] 5.2219159 1.8229953
#> [109,] 5.5105288 1.5343823
#> [110,] 7.0449111 0.0000000
#> [111,] 4.7779250 2.2669862
#> [112,] 5.1187209 1.9261903
#> [113,] 5.0640837 1.9808275
#> [114,] 5.1187209 1.9261903
#> [115,] 5.8214186 1.2234925
#> [116,] 4.4280277 2.6168834
#> [117,] 5.1220684 1.9228428
#> [118,] 5.8580183 1.1868928
#> [119,] 5.2186313 1.8262799
#> [120,] 3.6988351 3.3460760
#> [121,] 4.9116226 2.1332885
#> [122,] 4.9371848 2.1077263
#> [123,] 5.6149205 1.4299906
#> [124,] 5.5652289 1.4796822
#> [125,] 5.7007849 1.3441263
#> [126,] 4.8658157 2.1790954
#> [127,] 5.8214186 1.2234925
#> [128,] 5.6428002 1.4021110
#> [129,] 4.9303231 2.1145880
#> [130,] 5.5783837 1.4665275
#> [131,] 5.5783837 1.4665275
#> [132,] 5.6696206 1.3752905
#> [133,] 4.8970709 2.1478402
#> [134,] 4.9883078 2.0566033
#> [135,] 6.3143337 0.7305775
#> [136,] 6.3143337 0.7305775
#> [137,] 6.3143337 0.7305775
#> [138,] 5.5652289 1.4796822
#> [139,] 4.6188332 2.4260779
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.20344831 0.79655169
#>   [2,] 0.30024340 0.69975660
#>   [3,] 0.32774840 0.67225160
#>   [4,] 0.21780010 0.78219990
#>   [5,] 0.34483935 0.65516065
#>   [6,] 0.30799896 0.69200104
#>   [7,] 0.28797008 0.71202992
#>   [8,] 0.40525981 0.59474019
#>   [9,] 0.19902465 0.80097535
#>  [10,] 0.21826633 0.78173367
#>  [11,] 0.19598912 0.80401088
#>  [12,] 0.21826633 0.78173367
#>  [13,] 0.27851712 0.72148288
#>  [14,] 0.27332192 0.72667808
#>  [15,] 0.29192750 0.70807250
#>  [16,] 0.40470888 0.59529112
#>  [17,] 0.27774558 0.72225442
#>  [18,] 0.34879677 0.65120323
#>  [19,] 0.21780010 0.78219990
#>  [20,] 0.09927920 0.90072080
#>  [21,] 0.07696047 0.92303953
#>  [22,] 0.21780010 0.78219990
#>  [23,] 0.07696047 0.92303953
#>  [24,] 0.38268622 0.61731378
#>  [25,] 0.27294960 0.72705040
#>  [26,] 0.11852089 0.88147911
#>  [27,] 0.40525981 0.59474019
#>  [28,] 0.31373848 0.68626152
#>  [29,] 0.30581092 0.69418908
#>  [30,] 0.28797008 0.71202992
#>  [31,] 0.09670992 0.90329008
#>  [32,] 0.19598020 0.80401980
#>  [33,] 0.00000000 1.00000000
#>  [34,] 0.00000000 1.00000000
#>  [35,] 0.18180720 0.81819280
#>  [36,] 0.24509390 0.75490610
#>  [37,] 0.30015823 0.69984177
#>  [38,] 0.10370286 0.89629714
#>  [39,] 0.21523081 0.78476919
#>  [40,] 0.17923791 0.82076209
#>  [41,] 0.09927920 0.90072080
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.34437311 0.65562689
#>  [44,] 0.20353349 0.79646651
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.17367039 0.82632961
#>  [48,] 0.19645536 0.80354464
#>  [49,] 0.19548137 0.80451863
#>  [50,] 0.19902465 0.80097535
#>  [51,] 0.09927920 0.90072080
#>  [52,] 0.19902465 0.80097535
#>  [53,] 0.07696047 0.92303953
#>  [54,] 0.21003561 0.78996439
#>  [55,] 0.11852089 0.88147911
#>  [56,] 0.11852089 0.88147911
#>  [57,] 0.32205438 0.67794562
#>  [58,] 0.10446548 0.89553452
#>  [59,] 0.18699347 0.81300653
#>  [60,] 0.36323315 0.63676685
#>  [61,] 0.28070516 0.71929484
#>  [62,] 0.18699347 0.81300653
#>  [63,] 0.20374468 0.79625532
#>  [64,] 0.19645536 0.80354464
#>  [65,] 0.37711870 0.62288130
#>  [66,] 0.39589415 0.60410585
#>  [67,] 0.31497625 0.68502375
#>  [68,] 0.09670992 0.90329008
#>  [69,] 0.00000000 1.00000000
#>  [70,] 0.07696047 0.92303953
#>  [71,] 0.00000000 1.00000000
#>  [72,] 0.09974544 0.90025456
#>  [73,] 0.69650988 0.30349012
#>  [74,] 0.79183166 0.20816834
#>  [75,] 0.72225442 0.27774558
#>  [76,] 0.72225442 0.27774558
#>  [77,] 0.72199064 0.27800936
#>  [78,] 0.54525213 0.45474787
#>  [79,] 0.71448994 0.28551006
#>  [80,] 0.71402370 0.28597630
#>  [81,] 0.63250016 0.36749984
#>  [82,] 0.60567259 0.39432741
#>  [83,] 0.69938427 0.30061573
#>  [84,] 0.68216682 0.31783318
#>  [85,] 0.79646651 0.20353349
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.91747200 0.08252800
#>  [88,] 0.62221260 0.37778740
#>  [89,] 1.00000000 0.00000000
#>  [90,] 0.90848528 0.09151472
#>  [91,] 1.00000000 0.00000000
#>  [92,] 0.90072080 0.09927920
#>  [93,] 0.92303953 0.07696047
#>  [94,] 0.72225442 0.27774558
#>  [95,] 0.82595728 0.17404272
#>  [96,] 0.82595728 0.17404272
#>  [97,] 0.72705933 0.27294067
#>  [98,] 0.71929484 0.28070516
#>  [99,] 0.69672107 0.30327893
#> [100,] 0.70081577 0.29918423
#> [101,] 0.73604605 0.26395395
#> [102,] 0.72149181 0.27850819
#> [103,] 0.70930367 0.29069633
#> [104,] 0.92303953 0.07696047
#> [105,] 1.00000000 0.00000000
#> [106,] 0.74899681 0.25100319
#> [107,] 0.79089899 0.20910101
#> [108,] 0.74123232 0.25876768
#> [109,] 0.78219990 0.21780010
#> [110,] 1.00000000 0.00000000
#> [111,] 0.67820940 0.32179060
#> [112,] 0.72658416 0.27341584
#> [113,] 0.71882861 0.28117139
#> [114,] 0.72658416 0.27341584
#> [115,] 0.82632961 0.17367039
#> [116,] 0.62854273 0.37145727
#> [117,] 0.72705933 0.27294067
#> [118,] 0.83152480 0.16847520
#> [119,] 0.74076609 0.25923391
#> [120,] 0.52503645 0.47496355
#> [121,] 0.69718731 0.30281269
#> [122,] 0.70081577 0.29918423
#> [123,] 0.79701793 0.20298207
#> [124,] 0.78996439 0.21003561
#> [125,] 0.80920607 0.19079393
#> [126,] 0.69068518 0.30931482
#> [127,] 0.82632961 0.17367039
#> [128,] 0.80097535 0.19902465
#> [129,] 0.69984177 0.30015823
#> [130,] 0.79183166 0.20816834
#> [131,] 0.79183166 0.20816834
#> [132,] 0.80478241 0.19521759
#> [133,] 0.69512174 0.30487826
#> [134,] 0.70807250 0.29192750
#> [135,] 0.89629714 0.10370286
#> [136,] 0.89629714 0.10370286
#> [137,] 0.89629714 0.10370286
#> [138,] 0.78996439 0.21003561
#> [139,] 0.65562689 0.34437311
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.0317231  7.9910538 14.5936118  2.3924611  0.0000000  0.0000000  2.0764552 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  2.0615320  1.0094208  0.0000000  0.0000000  1.3723232  4.1200831 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  6.6127209  0.0000000  2.6094113  0.0000000  0.0000000  0.0000000  1.2649790 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  1.7272972  1.2421421  0.0000000  3.4788947  2.2448996  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.1104098  5.8421348  4.8281985  0.0000000  0.0000000  2.0842656  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  2.6273179  0.0000000  0.0000000  2.8780320  2.3689392  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  8.9277669  2.2539626  0.0000000  0.8201093  0.0000000  3.0610365  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  2.5869532  1.0875041  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  2.7506368  0.0000000  0.9437240 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 67 72 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V48        V10        V22        V36        V37        V21 
#> 14.5936118  8.9277669  7.9910538  6.6127209  5.8421348  4.8281985  4.1200831 
#>        V31        V52        V45         V7        V42        V24        V54 
#>  3.4788947  3.0610365  2.8780320  2.7506368  2.6273179  2.6094113  2.5869532 
#>        V12        V46        V49        V32         V4        V15        V17 
#>  2.3924611  2.3689392  2.2539626  2.2448996  2.0842656  2.0764552  2.0615320 
#>        V29        V20        V28         V3        V35        V55         V1 
#>  1.7272972  1.3723232  1.2649790  1.2421421  1.1104098  1.0875041  1.0317231 
#>        V18         V9        V50        V13        V14        V16        V19 
#>  1.0094208  0.9437240  0.8201093  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V2        V23        V25        V26        V27        V30        V33 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V34        V38        V39        V40        V41        V43        V44 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V47         V5        V51        V53        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6        V60         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913