Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V11>=0.2148 80 14 M (0.82500000 0.17500000)  
#>      4) V6< 0.1809 72  7 M (0.90277778 0.09722222) *
#>      5) V6>=0.1809 8  1 R (0.12500000 0.87500000) *
#>    3) V11< 0.2148 59 12 R (0.20338983 0.79661017)  
#>      6) V31< 0.4362 21 10 M (0.52380952 0.47619048)  
#>       12) V23>=0.8046 7  0 M (1.00000000 0.00000000) *
#>       13) V23< 0.8046 14  4 R (0.28571429 0.71428571) *
#>      7) V31>=0.4362 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V12>=0.1469 105 35 M (0.66666667 0.33333333)  
#>      4) V51>=0.01215 67 10 M (0.85074627 0.14925373)  
#>        8) V25>=0.4685 54  3 M (0.94444444 0.05555556) *
#>        9) V25< 0.4685 13  6 R (0.46153846 0.53846154) *
#>      5) V51< 0.01215 38 13 R (0.34210526 0.65789474)  
#>       10) V2< 0.02095 13  2 M (0.84615385 0.15384615) *
#>       11) V2>=0.02095 25  2 R (0.08000000 0.92000000) *
#>    3) V12< 0.1469 34  3 R (0.08823529 0.91176471) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V12>=0.22505 88 20 M (0.77272727 0.22727273)  
#>     4) V48>=0.04615 74 10 M (0.86486486 0.13513514)  
#>       8) V19< 0.8773 65  4 M (0.93846154 0.06153846) *
#>       9) V19>=0.8773 9  3 R (0.33333333 0.66666667) *
#>     5) V48< 0.04615 14  4 R (0.28571429 0.71428571) *
#>   3) V12< 0.22505 51 13 R (0.25490196 0.74509804)  
#>     6) V31< 0.34415 19  6 M (0.68421053 0.31578947) *
#>     7) V31>=0.34415 32  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V10>=0.16315 88 19 M (0.78409091 0.21590909)  
#>      4) V39>=0.0814 81 12 M (0.85185185 0.14814815)  
#>        8) V58>=0.00495 52  1 M (0.98076923 0.01923077) *
#>        9) V58< 0.00495 29 11 M (0.62068966 0.37931034)  
#>         18) V60< 0.00435 15  0 M (1.00000000 0.00000000) *
#>         19) V60>=0.00435 14  3 R (0.21428571 0.78571429) *
#>      5) V39< 0.0814 7  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.16315 51 13 R (0.25490196 0.74509804)  
#>      6) V31< 0.3755 21  8 M (0.61904762 0.38095238)  
#>       12) V12>=0.1671 11  0 M (1.00000000 0.00000000) *
#>       13) V12< 0.1671 10  2 R (0.20000000 0.80000000) *
#>      7) V31>=0.3755 30  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V11>=0.21705 76 27 M (0.64473684 0.35526316)  
#>      4) V51>=0.01375 44  6 M (0.86363636 0.13636364)  
#>        8) V34>=0.11155 37  0 M (1.00000000 0.00000000) *
#>        9) V34< 0.11155 7  1 R (0.14285714 0.85714286) *
#>      5) V51< 0.01375 32 11 R (0.34375000 0.65625000)  
#>       10) V16< 0.2119 7  0 M (1.00000000 0.00000000) *
#>       11) V16>=0.2119 25  4 R (0.16000000 0.84000000)  
#>         22) V35< 0.3663 7  3 M (0.57142857 0.42857143) *
#>         23) V35>=0.3663 18  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.21705 63  9 R (0.14285714 0.85714286)  
#>      6) V47>=0.16355 8  1 M (0.87500000 0.12500000) *
#>      7) V47< 0.16355 55  2 R (0.03636364 0.96363636) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 64 M (0.53956835 0.46043165)  
#>   2) V36< 0.40615 71 16 M (0.77464789 0.22535211)  
#>     4) V9>=0.10995 63  8 M (0.87301587 0.12698413)  
#>       8) V47>=0.05595 56  2 M (0.96428571 0.03571429) *
#>       9) V47< 0.05595 7  1 R (0.14285714 0.85714286) *
#>     5) V9< 0.10995 8  0 R (0.00000000 1.00000000) *
#>   3) V36>=0.40615 68 20 R (0.29411765 0.70588235)  
#>     6) V45>=0.26515 17  2 M (0.88235294 0.11764706) *
#>     7) V45< 0.26515 51  5 R (0.09803922 0.90196078) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V5>=0.0394 106 41 M (0.61320755 0.38679245)  
#>      4) V37< 0.47995 66 13 M (0.80303030 0.19696970)  
#>        8) V42>=0.0851 54  3 M (0.94444444 0.05555556) *
#>        9) V42< 0.0851 12  2 R (0.16666667 0.83333333) *
#>      5) V37>=0.47995 40 12 R (0.30000000 0.70000000)  
#>       10) V36>=0.75575 15  3 M (0.80000000 0.20000000) *
#>       11) V36< 0.75575 25  0 R (0.00000000 1.00000000) *
#>    3) V5< 0.0394 33  4 R (0.12121212 0.87878788)  
#>      6) V14< 0.1101 7  3 M (0.57142857 0.42857143) *
#>      7) V14>=0.1101 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 R (0.41007194 0.58992806)  
#>    2) V54>=0.02075 21  0 M (1.00000000 0.00000000) *
#>    3) V54< 0.02075 118 36 R (0.30508475 0.69491525)  
#>      6) V51>=0.01625 45 18 M (0.60000000 0.40000000)  
#>       12) V35< 0.669 30  4 M (0.86666667 0.13333333) *
#>       13) V35>=0.669 15  1 R (0.06666667 0.93333333) *
#>      7) V51< 0.01625 73  9 R (0.12328767 0.87671233)  
#>       14) V59< 0.0027 8  2 M (0.75000000 0.25000000) *
#>       15) V59>=0.0027 65  3 R (0.04615385 0.95384615) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 65 R (0.46762590 0.53237410)  
#>   2) V4>=0.03755 80 22 M (0.72500000 0.27500000)  
#>     4) V12>=0.1677 62  7 M (0.88709677 0.11290323)  
#>       8) V50>=0.0126 51  0 M (1.00000000 0.00000000) *
#>       9) V50< 0.0126 11  4 R (0.36363636 0.63636364) *
#>     5) V12< 0.1677 18  3 R (0.16666667 0.83333333) *
#>   3) V4< 0.03755 59  7 R (0.11864407 0.88135593)  
#>     6) V45>=0.3586 7  2 M (0.71428571 0.28571429) *
#>     7) V45< 0.3586 52  2 R (0.03846154 0.96153846) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V21>=0.58055 74 14 M (0.81081081 0.18918919)  
#>      4) V40>=0.1373 66  8 M (0.87878788 0.12121212)  
#>        8) V21< 0.97905 57  2 M (0.96491228 0.03508772) *
#>        9) V21>=0.97905 9  3 R (0.33333333 0.66666667) *
#>      5) V40< 0.1373 8  2 R (0.25000000 0.75000000) *
#>    3) V21< 0.58055 65 19 R (0.29230769 0.70769231)  
#>      6) V12>=0.2438 26 10 M (0.61538462 0.38461538)  
#>       12) V12< 0.35845 17  1 M (0.94117647 0.05882353) *
#>       13) V12>=0.35845 9  0 R (0.00000000 1.00000000) *
#>      7) V12< 0.2438 39  3 R (0.07692308 0.92307692) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.6885190 0.6037182 0.8622371 0.9915472 0.9089388 0.8186523
#>  [8] 0.6861218 0.6013833 0.7351953
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.9731267 4.7066257
#>   [2,] 1.5952655 6.0844869
#>   [3,] 2.5063712 5.1733812
#>   [4,] 2.2790524 5.4006999
#>   [5,] 1.6020915 6.0776609
#>   [6,] 2.2835542 5.3961981
#>   [7,] 3.2462255 4.4335268
#>   [8,] 1.4223705 6.2573818
#>   [9,] 0.0000000 7.6797523
#>  [10,] 2.0732792 5.6064731
#>  [11,] 2.0673387 5.6124137
#>  [12,] 2.3763697 5.3033826
#>  [13,] 1.9783591 5.7013932
#>  [14,] 0.0000000 7.6797523
#>  [15,] 1.8537843 5.8259680
#>  [16,] 2.1988157 5.4809366
#>  [17,] 3.0680468 4.6117056
#>  [18,] 0.8186523 6.8611000
#>  [19,] 1.7275911 5.9521612
#>  [20,] 1.3871574 6.2925949
#>  [21,] 2.3763697 5.3033826
#>  [22,] 1.2051016 6.4746507
#>  [23,] 0.8186523 6.8611000
#>  [24,] 1.4213171 6.2584352
#>  [25,] 2.7112195 4.9685328
#>  [26,] 0.7351953 6.9445571
#>  [27,] 0.8186523 6.8611000
#>  [28,] 0.8186523 6.8611000
#>  [29,] 1.5483589 6.1313934
#>  [30,] 0.6885190 6.9912333
#>  [31,] 0.0000000 7.6797523
#>  [32,] 1.2922373 6.3875150
#>  [33,] 2.5018694 5.1778829
#>  [34,] 2.1988157 5.4809366
#>  [35,] 1.6456762 6.0340761
#>  [36,] 2.2478523 5.4319000
#>  [37,] 2.0274325 5.6523198
#>  [38,] 1.3389135 6.3408388
#>  [39,] 0.0000000 7.6797523
#>  [40,] 1.5071714 6.1725810
#>  [41,] 0.6885190 6.9912333
#>  [42,] 1.4213171 6.2584352
#>  [43,] 1.4213171 6.2584352
#>  [44,] 0.8622371 6.8175153
#>  [45,] 1.4213171 6.2584352
#>  [46,] 1.5071714 6.1725810
#>  [47,] 2.1575658 5.5221865
#>  [48,] 0.9089388 6.7708135
#>  [49,] 0.6861218 6.9936305
#>  [50,] 2.1084924 5.5712599
#>  [51,] 1.4223705 6.2573818
#>  [52,] 1.4223705 6.2573818
#>  [53,] 2.2047563 5.4749961
#>  [54,] 0.9089388 6.7708135
#>  [55,] 1.3871574 6.2925949
#>  [56,] 2.5101816 5.1695707
#>  [57,] 2.6839252 4.9958271
#>  [58,] 1.4695610 6.2101913
#>  [59,] 2.4635054 5.2162469
#>  [60,] 2.3317981 5.3479543
#>  [61,] 1.5538476 6.1259048
#>  [62,] 1.5974579 6.0822945
#>  [63,] 1.5186344 6.1611179
#>  [64,] 1.7275911 5.9521612
#>  [65,] 0.0000000 7.6797523
#>  [66,] 5.9873744 1.6923780
#>  [67,] 6.2584352 1.4213171
#>  [68,] 5.4428744 2.2368780
#>  [69,] 5.6546547 2.0250977
#>  [70,] 6.1725810 1.5071714
#>  [71,] 5.7013932 1.9783591
#>  [72,] 4.8795901 2.8001623
#>  [73,] 5.0907472 2.5890051
#>  [74,] 5.0931444 2.5866079
#>  [75,] 5.1778829 2.5018694
#>  [76,] 5.5573997 2.1223527
#>  [77,] 6.1694302 1.5103222
#>  [78,] 5.9530098 1.7267425
#>  [79,] 6.6882051 0.9915472
#>  [80,] 6.0020832 1.6776691
#>  [81,] 6.0020832 1.6776691
#>  [82,] 6.0844869 1.5952655
#>  [83,] 7.6797523 0.0000000
#>  [84,] 6.4746507 1.2051016
#>  [85,] 6.3898499 1.2899024
#>  [86,] 6.2573818 1.4223705
#>  [87,] 7.0760341 0.6037182
#>  [88,] 4.5636600 3.1160923
#>  [89,] 5.4373857 2.2423666
#>  [90,] 6.9912333 0.6885190
#>  [91,] 6.3898499 1.2899024
#>  [92,] 6.3051114 1.3746409
#>  [93,] 5.9521612 1.7275911
#>  [94,] 6.8963132 0.7834391
#>  [95,] 6.8611000 0.8186523
#>  [96,] 5.3479543 2.3317981
#>  [97,] 6.1289962 1.5507561
#>  [98,] 7.6797523 0.0000000
#>  [99,] 5.1778829 2.5018694
#> [100,] 5.6546547 2.0250977
#> [101,] 4.9698766 2.7098757
#> [102,] 5.3983027 2.2814496
#> [103,] 6.8611000 0.8186523
#> [104,] 6.9912333 0.6885190
#> [105,] 6.9936305 0.6861218
#> [106,] 6.8611000 0.8186523
#> [107,] 5.7394555 1.9402968
#> [108,] 6.0868217 1.5929306
#> [109,] 6.9936305 0.6861218
#> [110,] 6.7708135 0.9089388
#> [111,] 4.2674973 3.4122550
#> [112,] 5.3455571 2.3341953
#> [113,] 4.7969817 2.8827707
#> [114,] 6.3899122 1.2898401
#> [115,] 5.2660393 2.4137130
#> [116,] 6.3922471 1.2875052
#> [117,] 6.1749782 1.5047742
#> [118,] 6.9936305 0.6861218
#> [119,] 4.8795901 2.8001623
#> [120,] 5.7037281 1.9760242
#> [121,] 6.8963132 0.7834391
#> [122,] 7.6797523 0.0000000
#> [123,] 7.6797523 0.0000000
#> [124,] 7.6797523 0.0000000
#> [125,] 6.2584352 1.4213171
#> [126,] 6.9936305 0.6861218
#> [127,] 6.2597167 1.4200357
#> [128,] 7.0783690 0.6013833
#> [129,] 7.0783690 0.6013833
#> [130,] 7.0783690 0.6013833
#> [131,] 6.0823200 1.5974323
#> [132,] 6.3431737 1.3365786
#> [133,] 6.0823200 1.5974323
#> [134,] 6.2597167 1.4200357
#> [135,] 6.2161319 1.4636204
#> [136,] 6.9445571 0.7351953
#> [137,] 6.1749782 1.5047742
#> [138,] 7.0783690 0.6013833
#> [139,] 5.6124137 2.0673387
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.38713835 0.61286165
#>   [2,] 0.20772356 0.79227644
#>   [3,] 0.32636094 0.67363906
#>   [4,] 0.29676119 0.70323881
#>   [5,] 0.20861239 0.79138761
#>   [6,] 0.29734737 0.70265263
#>   [7,] 0.42269925 0.57730075
#>   [8,] 0.18521047 0.81478953
#>   [9,] 0.00000000 1.00000000
#>  [10,] 0.26996694 0.73003306
#>  [11,] 0.26919340 0.73080660
#>  [12,] 0.30943312 0.69056688
#>  [13,] 0.25760715 0.74239285
#>  [14,] 0.00000000 1.00000000
#>  [15,] 0.24138595 0.75861405
#>  [16,] 0.28631336 0.71368664
#>  [17,] 0.39949814 0.60050186
#>  [18,] 0.10659879 0.89340121
#>  [19,] 0.22495402 0.77504598
#>  [20,] 0.18062528 0.81937472
#>  [21,] 0.30943312 0.69056688
#>  [22,] 0.15691933 0.84308067
#>  [23,] 0.10659879 0.89340121
#>  [24,] 0.18507330 0.81492670
#>  [25,] 0.35303476 0.64696524
#>  [26,] 0.09573164 0.90426836
#>  [27,] 0.10659879 0.89340121
#>  [28,] 0.10659879 0.89340121
#>  [29,] 0.20161574 0.79838426
#>  [30,] 0.08965381 0.91034619
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.16826549 0.83173451
#>  [33,] 0.32577475 0.67422525
#>  [34,] 0.28631336 0.71368664
#>  [35,] 0.21428767 0.78571233
#>  [36,] 0.29269854 0.70730146
#>  [37,] 0.26399713 0.73600287
#>  [38,] 0.17434332 0.82565668
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.19625260 0.80374740
#>  [41,] 0.08965381 0.91034619
#>  [42,] 0.18507330 0.81492670
#>  [43,] 0.18507330 0.81492670
#>  [44,] 0.11227407 0.88772593
#>  [45,] 0.18507330 0.81492670
#>  [46,] 0.19625260 0.80374740
#>  [47,] 0.28094211 0.71905789
#>  [48,] 0.11835523 0.88164477
#>  [49,] 0.08934166 0.91065834
#>  [50,] 0.27455214 0.72544786
#>  [51,] 0.18521047 0.81478953
#>  [52,] 0.18521047 0.81478953
#>  [53,] 0.28708690 0.71291310
#>  [54,] 0.11835523 0.88164477
#>  [55,] 0.18062528 0.81937472
#>  [56,] 0.32685711 0.67314289
#>  [57,] 0.34948070 0.65051930
#>  [58,] 0.19135526 0.80864474
#>  [59,] 0.32077928 0.67922072
#>  [60,] 0.30362933 0.69637067
#>  [61,] 0.20233043 0.79766957
#>  [62,] 0.20800903 0.79199097
#>  [63,] 0.19774523 0.80225477
#>  [64,] 0.22495402 0.77504598
#>  [65,] 0.00000000 1.00000000
#>  [66,] 0.77963118 0.22036882
#>  [67,] 0.81492670 0.18507330
#>  [68,] 0.70873045 0.29126955
#>  [69,] 0.73630691 0.26369309
#>  [70,] 0.80374740 0.19625260
#>  [71,] 0.74239285 0.25760715
#>  [72,] 0.63538378 0.36461622
#>  [73,] 0.66287909 0.33712091
#>  [74,] 0.66319123 0.33680877
#>  [75,] 0.67422525 0.32577475
#>  [76,] 0.72364309 0.27635691
#>  [77,] 0.80333713 0.19666287
#>  [78,] 0.77515649 0.22484351
#>  [79,] 0.87088812 0.12911188
#>  [80,] 0.78154646 0.21845354
#>  [81,] 0.78154646 0.21845354
#>  [82,] 0.79227644 0.20772356
#>  [83,] 1.00000000 0.00000000
#>  [84,] 0.84308067 0.15691933
#>  [85,] 0.83203854 0.16796146
#>  [86,] 0.81478953 0.18521047
#>  [87,] 0.92138832 0.07861168
#>  [88,] 0.59424573 0.40575427
#>  [89,] 0.70801576 0.29198424
#>  [90,] 0.91034619 0.08965381
#>  [91,] 0.83203854 0.16796146
#>  [92,] 0.82100453 0.17899547
#>  [93,] 0.77504598 0.22495402
#>  [94,] 0.89798640 0.10201360
#>  [95,] 0.89340121 0.10659879
#>  [96,] 0.69637067 0.30362933
#>  [97,] 0.79807212 0.20192788
#>  [98,] 1.00000000 0.00000000
#>  [99,] 0.67422525 0.32577475
#> [100,] 0.73630691 0.26369309
#> [101,] 0.64714022 0.35285978
#> [102,] 0.70292667 0.29707333
#> [103,] 0.89340121 0.10659879
#> [104,] 0.91034619 0.08965381
#> [105,] 0.91065834 0.08934166
#> [106,] 0.89340121 0.10659879
#> [107,] 0.74734903 0.25265097
#> [108,] 0.79258048 0.20741952
#> [109,] 0.91065834 0.08934166
#> [110,] 0.88164477 0.11835523
#> [111,] 0.55568163 0.44431837
#> [112,] 0.69605852 0.30394148
#> [113,] 0.62462713 0.37537287
#> [114,] 0.83204666 0.16795334
#> [115,] 0.68570432 0.31429568
#> [116,] 0.83235069 0.16764931
#> [117,] 0.80405955 0.19594045
#> [118,] 0.91065834 0.08934166
#> [119,] 0.63538378 0.36461622
#> [120,] 0.74269688 0.25730312
#> [121,] 0.89798640 0.10201360
#> [122,] 1.00000000 0.00000000
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 0.81492670 0.18507330
#> [126,] 0.91065834 0.08934166
#> [127,] 0.81509356 0.18490644
#> [128,] 0.92169235 0.07830765
#> [129,] 0.92169235 0.07830765
#> [130,] 0.92169235 0.07830765
#> [131,] 0.79199429 0.20800571
#> [132,] 0.82596071 0.17403929
#> [133,] 0.79199429 0.20800571
#> [134,] 0.81509356 0.18490644
#> [135,] 0.80941828 0.19058172
#> [136,] 0.90426836 0.09573164
#> [137,] 0.80405955 0.19594045
#> [138,] 0.92169235 0.07830765
#> [139,] 0.73080660 0.26919340
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  3.9996973  9.6872276 13.1057449  0.0000000  0.7563481  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.9629123  0.0000000  0.0000000  0.8965904  1.7733259  0.0000000  4.6783699 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  0.9569667  0.0000000  0.8630666  0.0000000  0.0000000  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  5.1680839  0.0000000  0.0000000  2.1999245 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  3.0899576  6.2599293  2.6468340  0.0000000  2.0682169  3.8518974  1.0640305 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  2.4944941  0.0000000  0.0000000  4.5267188  0.0000000  4.4564633 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.2230247  0.0000000  2.5583943  1.1305093  6.9900141  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.0303466  0.0000000  0.0000000  0.0000000  1.0677777  1.2421615  1.7506111 
#>        V60         V7         V8         V9 
#>  1.9775075  0.0000000  0.0000000  2.5228532 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12        V11        V51        V36        V31        V21        V45 
#> 13.1057449  9.6872276  6.9900141  6.2599293  5.1680839  4.6783699  4.5267188 
#>        V47        V10         V4        V35        V54        V37         V5 
#>  4.4564633  3.9996973  3.8518974  3.0899576  3.0303466  2.6468340  2.5583943 
#>         V9        V42        V34        V39        V60        V16         V2 
#>  2.5228532  2.4944941  2.1999245  2.0682169  1.9775075  1.9629123  1.7733259 
#>         V6        V59        V48        V50        V58        V40        V23 
#>  1.7506111  1.2421615  1.2230247  1.1305093  1.0677777  1.0640305  0.9569667 
#>        V19        V25        V14         V1        V13        V15        V17 
#>  0.8965904  0.8630666  0.7563481  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V18        V20        V22        V24        V26        V27        V28 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V29         V3        V30        V32        V33        V38        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V43        V44        V46        V49        V52        V53        V55 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V56        V57         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986