Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V11>=0.17045 99 22 M (0.77777778 0.22222222)  
#>     4) V5< 0.15035 88 12 M (0.86363636 0.13636364)  
#>       8) V16< 0.838 79  5 M (0.93670886 0.06329114) *
#>       9) V16>=0.838 9  2 R (0.22222222 0.77777778) *
#>     5) V5>=0.15035 11  1 R (0.09090909 0.90909091) *
#>   3) V11< 0.17045 40  4 R (0.10000000 0.90000000)  
#>     6) V15>=0.33735 7  3 M (0.57142857 0.42857143) *
#>     7) V15< 0.33735 33  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V12>=0.16505 92 28 M (0.69565217 0.30434783)  
#>      4) V51>=0.01265 51  2 M (0.96078431 0.03921569) *
#>      5) V51< 0.01265 41 15 R (0.36585366 0.63414634)  
#>       10) V43>=0.2164 15  2 M (0.86666667 0.13333333) *
#>       11) V43< 0.2164 26  2 R (0.07692308 0.92307692) *
#>    3) V12< 0.16505 47  9 R (0.19148936 0.80851064)  
#>      6) V53>=0.0153 7  0 M (1.00000000 0.00000000) *
#>      7) V53< 0.0153 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V35< 0.244 50  9 M (0.82000000 0.18000000)  
#>      4) V17< 0.7199 40  2 M (0.95000000 0.05000000) *
#>      5) V17>=0.7199 10  3 R (0.30000000 0.70000000) *
#>    3) V35>=0.244 89 26 R (0.29213483 0.70786517)  
#>      6) V1>=0.04195 16  1 M (0.93750000 0.06250000) *
#>      7) V1< 0.04195 73 11 R (0.15068493 0.84931507)  
#>       14) V55< 0.0041 12  4 M (0.66666667 0.33333333) *
#>       15) V55>=0.0041 61  3 R (0.04918033 0.95081967) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.19825 89 24 M (0.73033708 0.26966292)  
#>      4) V18< 0.63795 54  3 M (0.94444444 0.05555556) *
#>      5) V18>=0.63795 35 14 R (0.40000000 0.60000000)  
#>       10) V19>=0.86315 13  1 M (0.92307692 0.07692308) *
#>       11) V19< 0.86315 22  2 R (0.09090909 0.90909091) *
#>    3) V11< 0.19825 50  9 R (0.18000000 0.82000000)  
#>      6) V9>=0.1337 9  3 M (0.66666667 0.33333333) *
#>      7) V9< 0.1337 41  3 R (0.07317073 0.92682927) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V49>=0.0595 45  7 M (0.84444444 0.15555556)  
#>      4) V12>=0.09485 38  2 M (0.94736842 0.05263158) *
#>      5) V12< 0.09485 7  2 R (0.28571429 0.71428571) *
#>    3) V49< 0.0595 94 30 R (0.31914894 0.68085106)  
#>      6) V32< 0.4972 58 29 M (0.50000000 0.50000000)  
#>       12) V12>=0.16675 36  8 M (0.77777778 0.22222222)  
#>         24) V33>=0.25535 27  2 M (0.92592593 0.07407407) *
#>         25) V33< 0.25535 9  3 R (0.33333333 0.66666667) *
#>       13) V12< 0.16675 22  1 R (0.04545455 0.95454545) *
#>      7) V32>=0.4972 36  1 R (0.02777778 0.97222222) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V11>=0.19825 83 27 M (0.67469880 0.32530120)  
#>      4) V7< 0.20865 70 16 M (0.77142857 0.22857143)  
#>        8) V52>=0.00885 53  5 M (0.90566038 0.09433962)  
#>         16) V58< 0.01155 44  0 M (1.00000000 0.00000000) *
#>         17) V58>=0.01155 9  4 R (0.44444444 0.55555556) *
#>        9) V52< 0.00885 17  6 R (0.35294118 0.64705882) *
#>      5) V7>=0.20865 13  2 R (0.15384615 0.84615385) *
#>    3) V11< 0.19825 56 13 R (0.23214286 0.76785714)  
#>      6) V59>=0.00775 16  5 M (0.68750000 0.31250000) *
#>      7) V59< 0.00775 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V20>=0.4977 65 17 M (0.73846154 0.26153846)  
#>      4) V51>=0.0108 40  3 M (0.92500000 0.07500000) *
#>      5) V51< 0.0108 25 11 R (0.44000000 0.56000000)  
#>       10) V2< 0.0217 10  0 M (1.00000000 0.00000000) *
#>       11) V2>=0.0217 15  1 R (0.06666667 0.93333333) *
#>    3) V20< 0.4977 74 13 R (0.17567568 0.82432432)  
#>      6) V12>=0.22545 25 12 M (0.52000000 0.48000000)  
#>       12) V13< 0.3394 13  0 M (1.00000000 0.00000000) *
#>       13) V13>=0.3394 12  0 R (0.00000000 1.00000000) *
#>      7) V12< 0.22545 49  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.4676259 0.5323741)  
#>    2) V31< 0.45155 67 20 M (0.7014925 0.2985075)  
#>      4) V45>=0.13865 32  1 M (0.9687500 0.0312500) *
#>      5) V45< 0.13865 35 16 R (0.4571429 0.5428571)  
#>       10) V15>=0.504 12  0 M (1.0000000 0.0000000) *
#>       11) V15< 0.504 23  4 R (0.1739130 0.8260870)  
#>         22) V55< 0.0063 7  3 M (0.5714286 0.4285714) *
#>         23) V55>=0.0063 16  0 R (0.0000000 1.0000000) *
#>    3) V31>=0.45155 72 18 R (0.2500000 0.7500000)  
#>      6) V44>=0.4021 8  0 M (1.0000000 0.0000000) *
#>      7) V44< 0.4021 64 10 R (0.1562500 0.8437500)  
#>       14) V25>=0.7817 17  7 M (0.5882353 0.4117647) *
#>       15) V25< 0.7817 47  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 59 R (0.42446043 0.57553957)  
#>   2) V10>=0.15435 79 30 M (0.62025316 0.37974684)  
#>     4) V37< 0.4787 63 16 M (0.74603175 0.25396825)  
#>       8) V15< 0.56165 46  3 M (0.93478261 0.06521739) *
#>       9) V15>=0.56165 17  4 R (0.23529412 0.76470588) *
#>     5) V37>=0.4787 16  2 R (0.12500000 0.87500000) *
#>   3) V10< 0.15435 60 10 R (0.16666667 0.83333333)  
#>     6) V45>=0.1728 10  2 M (0.80000000 0.20000000) *
#>     7) V45< 0.1728 50  2 R (0.04000000 0.96000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 R (0.47482014 0.52517986)  
#>   2) V10>=0.1506 78 20 M (0.74358974 0.25641026)  
#>     4) V18>=0.17275 66 11 M (0.83333333 0.16666667)  
#>       8) V30< 0.7443 58  5 M (0.91379310 0.08620690) *
#>       9) V30>=0.7443 8  2 R (0.25000000 0.75000000) *
#>     5) V18< 0.17275 12  3 R (0.25000000 0.75000000) *
#>   3) V10< 0.1506 61  8 R (0.13114754 0.86885246)  
#>     6) V33< 0.1844 8  2 M (0.75000000 0.25000000) *
#>     7) V33>=0.1844 53  2 R (0.03773585 0.96226415) *
#> 
#> 
#> $weights
#>  [1] 0.5445214 0.6167829 0.6247528 0.6111640 0.5790392 0.5292118 0.8981708
#>  [8] 0.7108756 0.7444411 0.5204023
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.3716538 4.0077081
#>   [2,] 2.1294487 4.2499132
#>   [3,] 1.1556854 5.2236765
#>   [4,] 1.9402631 4.4390988
#>   [5,] 1.0994415 5.2799204
#>   [6,] 1.7483492 4.6310127
#>   [7,] 2.4913138 3.8880481
#>   [8,] 0.0000000 6.3793619
#>   [9,] 1.5149537 4.8644083
#>  [10,] 1.4185731 4.9607888
#>  [11,] 2.4349572 3.9444047
#>  [12,] 2.8369933 3.5423686
#>  [13,] 1.9719040 4.4074579
#>  [14,] 0.7108756 5.6684863
#>  [15,] 2.7797970 3.5995649
#>  [16,] 1.7563191 4.6230428
#>  [17,] 1.7860571 4.5933048
#>  [18,] 0.5292118 5.8501501
#>  [19,] 1.8721799 4.5071820
#>  [20,] 1.9630945 4.4162674
#>  [21,] 2.4205288 3.9588331
#>  [22,] 1.7403435 4.6390184
#>  [23,] 1.8093648 4.5699971
#>  [24,] 1.4426922 4.9366697
#>  [25,] 2.7187456 3.6606163
#>  [26,] 1.1556854 5.2236765
#>  [27,] 0.6167829 5.7625790
#>  [28,] 0.7108756 5.6684863
#>  [29,] 0.6247528 5.7546091
#>  [30,] 1.8721799 4.5071820
#>  [31,] 1.9845285 4.3948334
#>  [32,] 1.7483492 4.6310127
#>  [33,] 2.4833439 3.8960180
#>  [34,] 0.0000000 6.3793619
#>  [35,] 1.3356284 5.0437335
#>  [36,] 1.8801498 4.4992121
#>  [37,] 2.4591890 3.9201729
#>  [38,] 0.7108756 5.6684863
#>  [39,] 1.2400874 5.1392745
#>  [40,] 1.3356284 5.0437335
#>  [41,] 1.1459947 5.2333672
#>  [42,] 1.1459947 5.2333672
#>  [43,] 1.3691939 5.0101680
#>  [44,] 0.5204023 5.8589596
#>  [45,] 1.3356284 5.0437335
#>  [46,] 1.1451551 5.2342068
#>  [47,] 2.6739702 3.7053917
#>  [48,] 2.1535678 4.2257941
#>  [49,] 1.2553970 5.1239649
#>  [50,] 2.1535678 4.2257941
#>  [51,] 2.5268241 3.8525378
#>  [52,] 2.6328954 3.7464665
#>  [53,] 1.7106055 4.6687564
#>  [54,] 1.7757994 4.6035625
#>  [55,] 1.7757994 4.6035625
#>  [56,] 1.7757994 4.6035625
#>  [57,] 1.2312779 5.1480840
#>  [58,] 3.1761799 3.2031820
#>  [59,] 1.8069861 4.5723758
#>  [60,] 1.1459947 5.2333672
#>  [61,] 2.3819115 3.9974504
#>  [62,] 4.4792354 1.9001265
#>  [63,] 5.3056287 1.0737332
#>  [64,] 5.3297477 1.0496142
#>  [65,] 3.9001962 2.4791657
#>  [66,] 4.4809562 1.8984057
#>  [67,] 5.6349208 0.7444411
#>  [68,] 3.9605539 2.4188080
#>  [69,] 3.7104436 2.6689183
#>  [70,] 3.9452443 2.4341176
#>  [71,] 3.9588331 2.4205288
#>  [72,] 4.3496248 2.0297372
#>  [73,] 3.9001962 2.4791657
#>  [74,] 5.8501501 0.5292118
#>  [75,] 5.7546091 0.6247528
#>  [76,] 3.8697922 2.5095697
#>  [77,] 5.7681979 0.6111640
#>  [78,] 5.7681979 0.6111640
#>  [79,] 5.0237568 1.3556051
#>  [80,] 5.7681979 0.6111640
#>  [81,] 5.7681979 0.6111640
#>  [82,] 4.5644059 1.8149560
#>  [83,] 4.4656466 1.9137153
#>  [84,] 5.1057090 1.2736529
#>  [85,] 4.4311289 1.9482330
#>  [86,] 3.8866074 2.4927545
#>  [87,] 4.4889262 1.8904357
#>  [88,] 5.1057090 1.2736529
#>  [89,] 5.1145185 1.2648434
#>  [90,] 3.3815147 2.9978472
#>  [91,] 4.0062675 2.3730944
#>  [92,] 5.8501501 0.5292118
#>  [93,] 4.3817496 1.9976123
#>  [94,] 3.7569968 2.6223651
#>  [95,] 5.1392745 1.2400874
#>  [96,] 5.3297477 1.0496142
#>  [97,] 5.3056287 1.0737332
#>  [98,] 5.8003227 0.5790392
#>  [99,] 4.7049950 1.6743669
#> [100,] 4.6543280 1.7250339
#> [101,] 5.1835399 1.1958220
#> [102,] 3.9121671 2.4671948
#> [103,] 3.6606163 2.7187456
#> [104,] 4.6543280 1.7250339
#> [105,] 5.1145185 1.2648434
#> [106,] 4.4897657 1.8895962
#> [107,] 5.1145185 1.2648434
#> [108,] 5.4811911 0.8981708
#> [109,] 5.8501501 0.5292118
#> [110,] 4.5145217 1.8648402
#> [111,] 4.0776818 2.3016801
#> [112,] 5.8589596 0.5204023
#> [113,] 5.3297477 1.0496142
#> [114,] 5.8589596 0.5204023
#> [115,] 4.0245234 2.3548385
#> [116,] 5.1835399 1.1958220
#> [117,] 5.8501501 0.5292118
#> [118,] 5.8501501 0.5292118
#> [119,] 4.6086144 1.7707475
#> [120,] 4.6543280 1.7250339
#> [121,] 6.3793619 0.0000000
#> [122,] 6.3793619 0.0000000
#> [123,] 6.3793619 0.0000000
#> [124,] 5.8501501 0.5292118
#> [125,] 5.1480840 1.2312779
#> [126,] 5.6684863 0.7108756
#> [127,] 5.6684863 0.7108756
#> [128,] 5.0894471 1.2899148
#> [129,] 5.6349208 0.7444411
#> [130,] 5.1392745 1.2400874
#> [131,] 5.3297477 1.0496142
#> [132,] 6.3793619 0.0000000
#> [133,] 6.3793619 0.0000000
#> [134,] 6.3793619 0.0000000
#> [135,] 6.3793619 0.0000000
#> [136,] 5.8003227 0.5790392
#> [137,] 6.3793619 0.0000000
#> [138,] 5.2333672 1.1459947
#> [139,] 5.2333672 1.1459947
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.37176975 0.62823025
#>   [2,] 0.33380278 0.66619722
#>   [3,] 0.18116004 0.81883996
#>   [4,] 0.30414689 0.69585311
#>   [5,] 0.17234349 0.82765651
#>   [6,] 0.27406334 0.72593666
#>   [7,] 0.39052712 0.60947288
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.23747730 0.76252270
#>  [10,] 0.22236913 0.77763087
#>  [11,] 0.38169290 0.61830710
#>  [12,] 0.44471427 0.55528573
#>  [13,] 0.30910678 0.69089322
#>  [14,] 0.11143365 0.88856635
#>  [15,] 0.43574845 0.56425155
#>  [16,] 0.27531267 0.72468733
#>  [17,] 0.27997425 0.72002575
#>  [18,] 0.08295686 0.91704314
#>  [19,] 0.29347447 0.70652553
#>  [20,] 0.30772585 0.69227415
#>  [21,] 0.37943118 0.62056882
#>  [22,] 0.27280839 0.72719161
#>  [23,] 0.28362787 0.71637213
#>  [24,] 0.22614992 0.77385008
#>  [25,] 0.42617830 0.57382170
#>  [26,] 0.18116004 0.81883996
#>  [27,] 0.09668410 0.90331590
#>  [28,] 0.11143365 0.88856635
#>  [29,] 0.09793343 0.90206657
#>  [30,] 0.29347447 0.70652553
#>  [31,] 0.31108574 0.68891426
#>  [32,] 0.27406334 0.72593666
#>  [33,] 0.38927779 0.61072221
#>  [34,] 0.00000000 1.00000000
#>  [35,] 0.20936708 0.79063292
#>  [36,] 0.29472381 0.70527619
#>  [37,] 0.38549138 0.61450862
#>  [38,] 0.11143365 0.88856635
#>  [39,] 0.19439051 0.80560949
#>  [40,] 0.20936708 0.79063292
#>  [41,] 0.17964096 0.82035904
#>  [42,] 0.17964096 0.82035904
#>  [43,] 0.21462865 0.78537135
#>  [44,] 0.08157592 0.91842408
#>  [45,] 0.20936708 0.79063292
#>  [46,] 0.17950935 0.82049065
#>  [47,] 0.41915950 0.58084050
#>  [48,] 0.33758358 0.66241642
#>  [49,] 0.19679038 0.80320962
#>  [50,] 0.33758358 0.66241642
#>  [51,] 0.39609356 0.60390644
#>  [52,] 0.41272081 0.58727919
#>  [53,] 0.26814681 0.73185319
#>  [54,] 0.27836630 0.72163370
#>  [55,] 0.27836630 0.72163370
#>  [56,] 0.27836630 0.72163370
#>  [57,] 0.19300958 0.80699042
#>  [58,] 0.49788364 0.50211636
#>  [59,] 0.28325499 0.71674501
#>  [60,] 0.17964096 0.82035904
#>  [61,] 0.37337770 0.62662230
#>  [62,] 0.70214474 0.29785526
#>  [63,] 0.83168642 0.16831358
#>  [64,] 0.83546722 0.16453278
#>  [65,] 0.61137717 0.38862283
#>  [66,] 0.70241449 0.29758551
#>  [67,] 0.88330478 0.11669522
#>  [68,] 0.62083857 0.37916143
#>  [69,] 0.58163241 0.41836759
#>  [70,] 0.61843870 0.38156130
#>  [71,] 0.62056882 0.37943118
#>  [72,] 0.68182756 0.31817244
#>  [73,] 0.61137717 0.38862283
#>  [74,] 0.91704314 0.08295686
#>  [75,] 0.90206657 0.09793343
#>  [76,] 0.60661117 0.39338883
#>  [77,] 0.90419668 0.09580332
#>  [78,] 0.90419668 0.09580332
#>  [79,] 0.78750146 0.21249854
#>  [80,] 0.90419668 0.09580332
#>  [81,] 0.90419668 0.09580332
#>  [82,] 0.71549568 0.28450432
#>  [83,] 0.70001462 0.29998538
#>  [84,] 0.80034792 0.19965208
#>  [85,] 0.69460377 0.30539623
#>  [86,] 0.60924705 0.39075295
#>  [87,] 0.70366382 0.29633618
#>  [88,] 0.80034792 0.19965208
#>  [89,] 0.80172885 0.19827115
#>  [90,] 0.53007099 0.46992901
#>  [91,] 0.62800442 0.37199558
#>  [92,] 0.91704314 0.08295686
#>  [93,] 0.68686330 0.31313670
#>  [94,] 0.58892987 0.41107013
#>  [95,] 0.80560949 0.19439051
#>  [96,] 0.83546722 0.16453278
#>  [97,] 0.83168642 0.16831358
#>  [98,] 0.90923243 0.09076757
#>  [99,] 0.73753379 0.26246621
#> [100,] 0.72959147 0.27040853
#> [101,] 0.81254833 0.18745167
#> [102,] 0.61325368 0.38674632
#> [103,] 0.57382170 0.42617830
#> [104,] 0.72959147 0.27040853
#> [105,] 0.80172885 0.19827115
#> [106,] 0.70379542 0.29620458
#> [107,] 0.80172885 0.19827115
#> [108,] 0.85920680 0.14079320
#> [109,] 0.91704314 0.08295686
#> [110,] 0.70767606 0.29232394
#> [111,] 0.63919900 0.36080100
#> [112,] 0.91842408 0.08157592
#> [113,] 0.83546722 0.16453278
#> [114,] 0.91842408 0.08157592
#> [115,] 0.63086613 0.36913387
#> [116,] 0.81254833 0.18745167
#> [117,] 0.91704314 0.08295686
#> [118,] 0.91704314 0.08295686
#> [119,] 0.72242561 0.27757439
#> [120,] 0.72959147 0.27040853
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 1.00000000 0.00000000
#> [124,] 0.91704314 0.08295686
#> [125,] 0.80699042 0.19300958
#> [126,] 0.88856635 0.11143365
#> [127,] 0.88856635 0.11143365
#> [128,] 0.79779877 0.20220123
#> [129,] 0.88330478 0.11669522
#> [130,] 0.80560949 0.19439051
#> [131,] 0.83546722 0.16453278
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 1.00000000 0.00000000
#> [135,] 1.00000000 0.00000000
#> [136,] 0.90923243 0.09076757
#> [137,] 1.00000000 0.00000000
#> [138,] 0.82035904 0.17964096
#> [139,] 0.82035904 0.17964096
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  3.1190079  7.3151211 10.1502142  8.9936976  3.4439244  0.0000000  5.7598809 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.3800671  1.2975825  3.4688723  2.1251408  2.8846546  6.0490289  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  0.0000000  0.0000000  1.8870033  0.0000000  0.0000000  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.9905784  3.0902766  1.7624367  1.9710336  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  3.4245974  0.0000000  2.2511346  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  2.2485127  2.2114073  4.1131046  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  0.0000000  2.9877963  1.9535193  0.0000000  5.0461676  1.2787201  2.0377447 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  2.1623537  0.0000000  0.0000000  0.7499183  1.5103999  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  1.3598555  0.0000000  0.9762470 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 78 61 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V12        V10        V20        V15        V51        V45 
#> 10.1502142  8.9936976  7.3151211  6.0490289  5.7598809  5.0461676  4.1131046 
#>        V18        V13        V35         V1        V31        V49         V2 
#>  3.4688723  3.4439244  3.4245974  3.1190079  3.0902766  2.9877963  2.8846546 
#>        V37        V43        V44        V55        V19        V53        V33 
#>  2.2511346  2.2485127  2.2114073  2.1623537  2.1251408  2.0377447  1.9710336 
#>         V5        V25        V32        V59        V16         V7        V17 
#>  1.9535193  1.8870033  1.7624367  1.5103999  1.3800671  1.3598555  1.2975825 
#>        V52        V30         V9        V58        V14        V21        V22 
#>  1.2787201  0.9905784  0.9762470  0.7499183  0.0000000  0.0000000  0.0000000 
#>        V23        V24        V26        V27        V28        V29         V3 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V34        V36        V38        V39         V4        V40        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V42        V46        V47        V48        V50        V54        V56 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V57         V6        V60         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1014493