Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V11>=0.14955 96 27 M (0.71875000 0.28125000)  
#>      4) V43>=0.1855 65  7 M (0.89230769 0.10769231) *
#>      5) V43< 0.1855 31 11 R (0.35483871 0.64516129)  
#>       10) V48>=0.0952 13  3 M (0.76923077 0.23076923) *
#>       11) V48< 0.0952 18  1 R (0.05555556 0.94444444) *
#>    3) V11< 0.14955 43  3 R (0.06976744 0.93023256) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V54>=0.0163 34  4 M (0.88235294 0.11764706) *
#>    3) V54< 0.0163 105 37 R (0.35238095 0.64761905)  
#>      6) V10>=0.19415 45 14 M (0.68888889 0.31111111)  
#>       12) V35< 0.2902 23  1 M (0.95652174 0.04347826) *
#>       13) V35>=0.2902 22  9 R (0.40909091 0.59090909)  
#>         26) V43>=0.3442 7  0 M (1.00000000 0.00000000) *
#>         27) V43< 0.3442 15  2 R (0.13333333 0.86666667) *
#>      7) V10< 0.19415 60  6 R (0.10000000 0.90000000)  
#>       14) V20>=0.86495 9  4 M (0.55555556 0.44444444) *
#>       15) V20< 0.86495 51  1 R (0.01960784 0.98039216) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.4244604 0.5755396)  
#>    2) V12>=0.15395 83 30 M (0.6385542 0.3614458)  
#>      4) V37< 0.48065 66 15 M (0.7727273 0.2272727)  
#>        8) V52>=0.00415 59  8 M (0.8644068 0.1355932)  
#>         16) V48>=0.0399 50  3 M (0.9400000 0.0600000) *
#>         17) V48< 0.0399 9  4 R (0.4444444 0.5555556) *
#>        9) V52< 0.00415 7  0 R (0.0000000 1.0000000) *
#>      5) V37>=0.48065 17  2 R (0.1176471 0.8823529) *
#>    3) V12< 0.15395 56  6 R (0.1071429 0.8928571)  
#>      6) V22>=0.90065 9  3 M (0.6666667 0.3333333) *
#>      7) V22< 0.90065 47  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V31< 0.37815 43  8 M (0.81395349 0.18604651)  
#>      4) V20>=0.6506 36  2 M (0.94444444 0.05555556) *
#>      5) V20< 0.6506 7  1 R (0.14285714 0.85714286) *
#>    3) V31>=0.37815 96 30 R (0.31250000 0.68750000)  
#>      6) V48>=0.1094 28  5 M (0.82142857 0.17857143)  
#>       12) V29>=0.52075 20  0 M (1.00000000 0.00000000) *
#>       13) V29< 0.52075 8  3 R (0.37500000 0.62500000) *
#>      7) V48< 0.1094 68  7 R (0.10294118 0.89705882)  
#>       14) V8>=0.1537 9  4 M (0.55555556 0.44444444) *
#>       15) V8< 0.1537 59  2 R (0.03389831 0.96610169) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V9>=0.145 79 23 M (0.70886076 0.29113924)  
#>      4) V6< 0.16325 59  7 M (0.88135593 0.11864407)  
#>        8) V37< 0.5221 48  1 M (0.97916667 0.02083333) *
#>        9) V37>=0.5221 11  5 R (0.45454545 0.54545455) *
#>      5) V6>=0.16325 20  4 R (0.20000000 0.80000000)  
#>       10) V29>=0.4865 7  3 M (0.57142857 0.42857143) *
#>       11) V29< 0.4865 13  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.145 60 13 R (0.21666667 0.78333333)  
#>      6) V54>=0.01845 7  1 M (0.85714286 0.14285714) *
#>      7) V54< 0.01845 53  7 R (0.13207547 0.86792453) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V42>=0.27145 59 18 M (0.69491525 0.30508475)  
#>      4) V13>=0.24965 35  2 M (0.94285714 0.05714286) *
#>      5) V13< 0.24965 24  8 R (0.33333333 0.66666667)  
#>       10) V4>=0.06205 8  1 M (0.87500000 0.12500000) *
#>       11) V4< 0.06205 16  1 R (0.06250000 0.93750000) *
#>    3) V42< 0.27145 80 18 R (0.22500000 0.77500000)  
#>      6) V28< 0.37365 7  1 M (0.85714286 0.14285714) *
#>      7) V28>=0.37365 73 12 R (0.16438356 0.83561644) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V51>=0.0118 83 20 M (0.75903614 0.24096386)  
#>      4) V11>=0.0894 75 12 M (0.84000000 0.16000000)  
#>        8) V36< 0.497 54  2 M (0.96296296 0.03703704) *
#>        9) V36>=0.497 21 10 M (0.52380952 0.47619048)  
#>         18) V45>=0.252 11  0 M (1.00000000 0.00000000) *
#>         19) V45< 0.252 10  0 R (0.00000000 1.00000000) *
#>      5) V11< 0.0894 8  0 R (0.00000000 1.00000000) *
#>    3) V51< 0.0118 56 12 R (0.21428571 0.78571429)  
#>      6) V23>=0.79225 18  6 M (0.66666667 0.33333333) *
#>      7) V23< 0.79225 38  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V44>=0.1959 51 12 M (0.76470588 0.23529412)  
#>      4) V34< 0.64645 41  4 M (0.90243902 0.09756098)  
#>        8) V14< 0.4086 34  0 M (1.00000000 0.00000000) *
#>        9) V14>=0.4086 7  3 R (0.42857143 0.57142857) *
#>      5) V34>=0.64645 10  2 R (0.20000000 0.80000000) *
#>    3) V44< 0.1959 88 29 R (0.32954545 0.67045455)  
#>      6) V7>=0.13465 34 12 M (0.64705882 0.35294118)  
#>       12) V38< 0.18445 18  0 M (1.00000000 0.00000000) *
#>       13) V38>=0.18445 16  4 R (0.25000000 0.75000000) *
#>      7) V7< 0.13465 54  7 R (0.12962963 0.87037037)  
#>       14) V4>=0.0477 13  6 M (0.53846154 0.46153846) *
#>       15) V4< 0.0477 41  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V52>=0.0082 88 25 M (0.71590909 0.28409091)  
#>      4) V48>=0.0719 52  5 M (0.90384615 0.09615385) *
#>      5) V48< 0.0719 36 16 R (0.44444444 0.55555556)  
#>       10) V27< 0.52375 13  2 M (0.84615385 0.15384615) *
#>       11) V27>=0.52375 23  5 R (0.21739130 0.78260870)  
#>         22) V27>=0.90985 7  2 M (0.71428571 0.28571429) *
#>         23) V27< 0.90985 16  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0082 51 11 R (0.21568627 0.78431373)  
#>      6) V30< 0.35115 7  0 M (1.00000000 0.00000000) *
#>      7) V30>=0.35115 44  4 R (0.09090909 0.90909091) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V11>=0.17885 89 31 M (0.65168539 0.34831461)  
#>      4) V3< 0.0367 42  2 M (0.95238095 0.04761905) *
#>      5) V3>=0.0367 47 18 R (0.38297872 0.61702128)  
#>       10) V56>=0.00925 14  1 M (0.92857143 0.07142857) *
#>       11) V56< 0.00925 33  5 R (0.15151515 0.84848485) *
#>    3) V11< 0.17885 50  8 R (0.16000000 0.84000000)  
#>      6) V60< 0.00175 7  2 M (0.71428571 0.28571429) *
#>      7) V60>=0.00175 43  3 R (0.06976744 0.93023256) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.7363982 0.6191396 0.6631237 0.5565737 0.6406376 0.8333288
#>  [8] 0.6590801 0.7104344 0.7131861
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 3.3546490 3.4885844
#>   [2,] 1.3695146 5.4737188
#>   [3,] 2.7894244 4.0538089
#>   [4,] 2.6999852 4.1432482
#>   [5,] 2.7290960 4.1141374
#>   [6,] 0.0000000 6.8432334
#>   [7,] 2.0034063 4.8398270
#>   [8,] 1.8603350 4.9828983
#>   [9,] 2.0876409 4.7555925
#>  [10,] 2.0101522 4.8330812
#>  [11,] 1.4217655 5.4214679
#>  [12,] 2.2328508 4.6103826
#>  [13,] 2.2068870 4.6363464
#>  [14,] 2.7255268 4.1177066
#>  [15,] 1.3763098 5.4669235
#>  [16,] 0.0000000 6.8432334
#>  [17,] 2.1555327 4.6877007
#>  [18,] 0.8333288 6.0099045
#>  [19,] 2.0642582 4.7789751
#>  [20,] 1.2697598 5.5734735
#>  [21,] 0.7131861 6.1300472
#>  [22,] 2.0034063 4.8398270
#>  [23,] 1.9886542 4.8545792
#>  [24,] 2.1059128 4.7373206
#>  [25,] 2.7621688 4.0810646
#>  [26,] 0.0000000 6.8432334
#>  [27,] 1.4924089 5.3508244
#>  [28,] 0.8333288 6.0099045
#>  [29,] 1.4217655 5.4214679
#>  [30,] 1.4236206 5.4196128
#>  [31,] 2.0326383 4.8105951
#>  [32,] 2.1852975 4.6579359
#>  [33,] 2.0110487 4.8321846
#>  [34,] 0.7104344 6.1327989
#>  [35,] 0.7104344 6.1327989
#>  [36,] 0.6191396 6.2240938
#>  [37,] 0.0000000 6.8432334
#>  [38,] 0.8333288 6.0099045
#>  [39,] 0.7113310 6.1319024
#>  [40,] 1.3323257 5.5109076
#>  [41,] 0.0000000 6.8432334
#>  [42,] 1.3899025 5.4533308
#>  [43,] 0.8333288 6.0099045
#>  [44,] 2.1030887 4.7401447
#>  [45,] 1.3899025 5.4533308
#>  [46,] 0.8333288 6.0099045
#>  [47,] 0.0000000 6.8432334
#>  [48,] 0.7363982 6.1068352
#>  [49,] 2.2578460 4.5853874
#>  [50,] 0.7131861 6.1300472
#>  [51,] 0.0000000 6.8432334
#>  [52,] 0.7131861 6.1300472
#>  [53,] 0.8333288 6.0099045
#>  [54,] 0.0000000 6.8432334
#>  [55,] 1.4964526 5.3467808
#>  [56,] 2.0659723 4.7772611
#>  [57,] 2.8993011 3.9439323
#>  [58,] 2.1115485 4.7316848
#>  [59,] 1.3222038 5.5210295
#>  [60,] 1.4524684 5.3907649
#>  [61,] 2.0687239 4.7745094
#>  [62,] 1.2679047 5.5753287
#>  [63,] 2.0651548 4.7780786
#>  [64,] 2.6234425 4.2197909
#>  [65,] 2.0624031 4.7808303
#>  [66,] 2.2550943 4.5881391
#>  [67,] 1.4217655 5.4214679
#>  [68,] 2.1108529 4.7323804
#>  [69,] 1.3899025 5.4533308
#>  [70,] 1.9954494 4.8477839
#>  [71,] 0.7104344 6.1327989
#>  [72,] 4.1672058 2.6760276
#>  [73,] 4.7548665 2.0883668
#>  [74,] 4.8536827 1.9895507
#>  [75,] 4.6543667 2.1888666
#>  [76,] 4.1517081 2.6915253
#>  [77,] 4.8621425 1.9810908
#>  [78,] 4.8702700 1.9729634
#>  [79,] 4.1990188 2.6442145
#>  [80,] 4.1517081 2.6915253
#>  [81,] 5.2967184 1.5465150
#>  [82,] 4.8911814 1.9520520
#>  [83,] 5.5435156 1.2997177
#>  [84,] 5.4661975 1.3770358
#>  [85,] 5.4921613 1.3510721
#>  [86,] 5.4661975 1.3770358
#>  [87,] 4.8730217 1.9702117
#>  [88,] 5.5435156 1.2997177
#>  [89,] 4.9869419 1.8562914
#>  [90,] 5.4709671 1.3722662
#>  [91,] 4.8303295 2.0129039
#>  [92,] 5.6275796 1.2156538
#>  [93,] 5.0084400 1.8347934
#>  [94,] 5.0084400 1.8347934
#>  [95,] 6.2240938 0.6191396
#>  [96,] 5.4669235 1.3763098
#>  [97,] 4.8518275 1.9914058
#>  [98,] 5.5394720 1.3037613
#>  [99,] 6.1300472 0.7131861
#> [100,] 6.1068352 0.7363982
#> [101,] 5.4477551 1.3954783
#> [102,] 4.6794339 2.1637994
#> [103,] 6.1300472 0.7131861
#> [104,] 5.4196128 1.4236206
#> [105,] 4.9203324 1.9229009
#> [106,] 4.1172844 2.7259489
#> [107,] 6.1841533 0.6590801
#> [108,] 4.9828983 1.8603350
#> [109,] 5.4912647 1.3519686
#> [110,] 4.6354498 2.2077836
#> [111,] 4.3678023 2.4754310
#> [112,] 5.0084400 1.8347934
#> [113,] 4.9543339 1.8888994
#> [114,] 5.0268825 1.8163509
#> [115,] 6.2025957 0.6406376
#> [116,] 5.4437114 1.3995219
#> [117,] 6.1068352 0.7363982
#> [118,] 4.2679982 2.5752352
#> [119,] 4.9243760 1.9188573
#> [120,] 5.4669235 1.3763098
#> [121,] 4.1699575 2.6732759
#> [122,] 5.5435156 1.2997177
#> [123,] 5.4894096 1.3538238
#> [124,] 6.1300472 0.7131861
#> [125,] 5.4709671 1.3722662
#> [126,] 6.1300472 0.7131861
#> [127,] 6.8432334 0.0000000
#> [128,] 5.4437114 1.3995219
#> [129,] 5.5502615 1.2929719
#> [130,] 5.5502615 1.2929719
#> [131,] 5.3964007 1.4468327
#> [132,] 6.2025957 0.6406376
#> [133,] 6.2025957 0.6406376
#> [134,] 4.8262859 2.0169475
#> [135,] 6.2025957 0.6406376
#> [136,] 5.5435156 1.2997177
#> [137,] 6.2025957 0.6406376
#> [138,] 4.3238182 2.5194151
#> [139,] 3.9929570 2.8502763
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.49021403 0.50978597
#>   [2,] 0.20012682 0.79987318
#>   [3,] 0.40761790 0.59238210
#>   [4,] 0.39454816 0.60545184
#>   [5,] 0.39880212 0.60119788
#>   [6,] 0.00000000 1.00000000
#>   [7,] 0.29275727 0.70724273
#>   [8,] 0.27185030 0.72814970
#>   [9,] 0.30506644 0.69493356
#>  [10,] 0.29374304 0.70625696
#>  [11,] 0.20776223 0.79223777
#>  [12,] 0.32628593 0.67371407
#>  [13,] 0.32249185 0.67750815
#>  [14,] 0.39828055 0.60171945
#>  [15,] 0.20111982 0.79888018
#>  [16,] 0.00000000 1.00000000
#>  [17,] 0.31498745 0.68501255
#>  [18,] 0.12177414 0.87822586
#>  [19,] 0.30164954 0.69835046
#>  [20,] 0.18554969 0.81445031
#>  [21,] 0.10421771 0.89578229
#>  [22,] 0.29275727 0.70724273
#>  [23,] 0.29060154 0.70939846
#>  [24,] 0.30773651 0.69226349
#>  [25,] 0.40363504 0.59636496
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.21808535 0.78191465
#>  [28,] 0.12177414 0.87822586
#>  [29,] 0.20776223 0.79223777
#>  [30,] 0.20803333 0.79196667
#>  [31,] 0.29702893 0.70297107
#>  [32,] 0.31933698 0.68066302
#>  [33,] 0.29387405 0.70612595
#>  [34,] 0.10381561 0.89618439
#>  [35,] 0.10381561 0.89618439
#>  [36,] 0.09047472 0.90952528
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.12177414 0.87822586
#>  [39,] 0.10394662 0.89605338
#>  [40,] 0.19469243 0.80530757
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.20310611 0.79689389
#>  [43,] 0.12177414 0.87822586
#>  [44,] 0.30732383 0.69267617
#>  [45,] 0.20310611 0.79689389
#>  [46,] 0.12177414 0.87822586
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.10760969 0.89239031
#>  [49,] 0.32993848 0.67006152
#>  [50,] 0.10421771 0.89578229
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.10421771 0.89578229
#>  [53,] 0.12177414 0.87822586
#>  [54,] 0.00000000 1.00000000
#>  [55,] 0.21867624 0.78132376
#>  [56,] 0.30190001 0.69809999
#>  [57,] 0.42367415 0.57632585
#>  [58,] 0.30856007 0.69143993
#>  [59,] 0.19321332 0.80678668
#>  [60,] 0.21224886 0.78775114
#>  [61,] 0.30230212 0.69769788
#>  [62,] 0.18527860 0.81472140
#>  [63,] 0.30178056 0.69821944
#>  [64,] 0.38336300 0.61663700
#>  [65,] 0.30137845 0.69862155
#>  [66,] 0.32953637 0.67046363
#>  [67,] 0.20776223 0.79223777
#>  [68,] 0.30845841 0.69154159
#>  [69,] 0.20310611 0.79689389
#>  [70,] 0.29159453 0.70840547
#>  [71,] 0.10381561 0.89618439
#>  [72,] 0.60895275 0.39104725
#>  [73,] 0.69482747 0.30517253
#>  [74,] 0.70926745 0.29073255
#>  [75,] 0.68014146 0.31985854
#>  [76,] 0.60668808 0.39331192
#>  [77,] 0.71050369 0.28949631
#>  [78,] 0.71169135 0.28830865
#>  [79,] 0.61360158 0.38639842
#>  [80,] 0.60668808 0.39331192
#>  [81,] 0.77400815 0.22599185
#>  [82,] 0.71474712 0.28525288
#>  [83,] 0.81007257 0.18992743
#>  [84,] 0.79877410 0.20122590
#>  [85,] 0.80256817 0.19743183
#>  [86,] 0.79877410 0.20122590
#>  [87,] 0.71209346 0.28790654
#>  [88,] 0.81007257 0.18992743
#>  [89,] 0.72874060 0.27125940
#>  [90,] 0.79947107 0.20052893
#>  [91,] 0.70585486 0.29414514
#>  [92,] 0.82235681 0.17764319
#>  [93,] 0.73188210 0.26811790
#>  [94,] 0.73188210 0.26811790
#>  [95,] 0.90952528 0.09047472
#>  [96,] 0.79888018 0.20111982
#>  [97,] 0.70899636 0.29100364
#>  [98,] 0.80948168 0.19051832
#>  [99,] 0.89578229 0.10421771
#> [100,] 0.89239031 0.10760969
#> [101,] 0.79607910 0.20392090
#> [102,] 0.68380452 0.31619548
#> [103,] 0.89578229 0.10421771
#> [104,] 0.79196667 0.20803333
#> [105,] 0.71900696 0.28099304
#> [106,] 0.60165776 0.39834224
#> [107,] 0.90368879 0.09631121
#> [108,] 0.72814970 0.27185030
#> [109,] 0.80243716 0.19756284
#> [110,] 0.67737713 0.32262287
#> [111,] 0.63826588 0.36173412
#> [112,] 0.73188210 0.26811790
#> [113,] 0.72397559 0.27602441
#> [114,] 0.73457709 0.26542291
#> [115,] 0.90638378 0.09361622
#> [116,] 0.79548821 0.20451179
#> [117,] 0.89239031 0.10760969
#> [118,] 0.62368152 0.37631848
#> [119,] 0.71959785 0.28040215
#> [120,] 0.79888018 0.20111982
#> [121,] 0.60935486 0.39064514
#> [122,] 0.81007257 0.18992743
#> [123,] 0.80216607 0.19783393
#> [124,] 0.89578229 0.10421771
#> [125,] 0.79947107 0.20052893
#> [126,] 0.89578229 0.10421771
#> [127,] 1.00000000 0.00000000
#> [128,] 0.79548821 0.20451179
#> [129,] 0.81105834 0.18894166
#> [130,] 0.81105834 0.18894166
#> [131,] 0.78857470 0.21142530
#> [132,] 0.90638378 0.09361622
#> [133,] 0.90638378 0.09361622
#> [134,] 0.70526396 0.29473604
#> [135,] 0.90638378 0.09361622
#> [136,] 0.81007257 0.18992743
#> [137,] 0.90638378 0.09361622
#> [138,] 0.63183849 0.36816151
#> [139,] 0.58348983 0.41651017
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  3.9602780 11.2580724  3.5259272  2.0435856  0.7534024  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  2.4817809  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  1.2535168  2.7281845  0.0000000  0.0000000  0.0000000  2.4712789  1.1843494 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  1.3913535  3.0929303  2.1384099  2.9864313  0.0000000  0.0000000  1.5766704 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.4965147  1.4654552  2.9927663  1.8938521  0.0000000  2.4977879  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  2.8970133  4.1931274  2.4302268  2.6324587  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  8.3667127  0.0000000  0.0000000  0.0000000  4.9869460  5.2073054  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  4.2947635  0.0000000  2.5528408  0.0000000  0.0000000  0.0000000  2.3275415 
#>        V60         V7         V8         V9 
#>  1.0755773  2.2202497  0.8498112  2.7728781 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 68 71 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V48        V52        V51        V54        V43        V10 
#> 11.2580724  8.3667127  5.2073054  4.9869460  4.2947635  4.1931274  3.9602780 
#>        V12         V3        V37        V31        V42         V9        V23 
#>  3.5259272  3.0929303  2.9927663  2.9864313  2.8970133  2.7728781  2.7281845 
#>        V45        V56         V4        V20        V27        V44         V6 
#>  2.6324587  2.5528408  2.4977879  2.4817809  2.4712789  2.4302268  2.3275415 
#>         V7        V30        V13        V38        V34        V35        V36 
#>  2.2202497  2.1384099  2.0435856  1.8938521  1.5766704  1.4965147  1.4654552 
#>        V29        V22        V28        V60         V8        V14         V1 
#>  1.3913535  1.2535168  1.1843494  1.0755773  0.8498112  0.7534024  0.0000000 
#>        V15        V16        V17        V18        V19         V2        V21 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V25        V26        V32        V33        V39        V40 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V41        V46        V47        V49         V5        V50        V53 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V55        V57        V58        V59 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841