Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.1707 89 18 M (0.79775281 0.20224719)  
#>      4) V10>=0.0929 82 12 M (0.85365854 0.14634146)  
#>        8) V16< 0.69445 70  5 M (0.92857143 0.07142857)  
#>         16) V40< 0.52875 63  1 M (0.98412698 0.01587302) *
#>         17) V40>=0.52875 7  3 R (0.42857143 0.57142857) *
#>        9) V16>=0.69445 12  5 R (0.41666667 0.58333333) *
#>      5) V10< 0.0929 7  1 R (0.14285714 0.85714286) *
#>    3) V11< 0.1707 50  3 R (0.06000000 0.94000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V10>=0.1631 81 17 M (0.79012346 0.20987654)  
#>      4) V36< 0.5309 59  5 M (0.91525424 0.08474576) *
#>      5) V36>=0.5309 22 10 R (0.45454545 0.54545455)  
#>       10) V22< 0.4944 10  0 M (1.00000000 0.00000000) *
#>       11) V22>=0.4944 12  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.1631 58 14 R (0.24137931 0.75862069)  
#>      6) V47>=0.1794 10  1 M (0.90000000 0.10000000) *
#>      7) V47< 0.1794 48  5 R (0.10416667 0.89583333)  
#>       14) V18>=0.69515 7  2 M (0.71428571 0.28571429) *
#>       15) V18< 0.69515 41  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 56 M (0.59712230 0.40287770)  
#>    2) V52>=0.00945 82 15 M (0.81707317 0.18292683)  
#>      4) V44>=0.17685 57  3 M (0.94736842 0.05263158) *
#>      5) V44< 0.17685 25 12 M (0.52000000 0.48000000)  
#>       10) V12>=0.23285 14  1 M (0.92857143 0.07142857) *
#>       11) V12< 0.23285 11  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.00945 57 16 R (0.28070175 0.71929825)  
#>      6) V45>=0.2368 11  1 M (0.90909091 0.09090909) *
#>      7) V45< 0.2368 46  6 R (0.13043478 0.86956522) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 R (0.48920863 0.51079137)  
#>   2) V11>=0.2026 78 22 M (0.71794872 0.28205128)  
#>     4) V17< 0.829 63  8 M (0.87301587 0.12698413) *
#>     5) V17>=0.829 15  1 R (0.06666667 0.93333333) *
#>   3) V11< 0.2026 61 12 R (0.19672131 0.80327869)  
#>     6) V22>=0.90175 12  4 M (0.66666667 0.33333333) *
#>     7) V22< 0.90175 49  4 R (0.08163265 0.91836735) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V35< 0.2208 44  6 M (0.86363636 0.13636364)  
#>      4) V17< 0.93465 36  0 M (1.00000000 0.00000000) *
#>      5) V17>=0.93465 8  2 R (0.25000000 0.75000000) *
#>    3) V35>=0.2208 95 33 R (0.34736842 0.65263158)  
#>      6) V11>=0.32 18  0 M (1.00000000 0.00000000) *
#>      7) V11< 0.32 77 15 R (0.19480519 0.80519481)  
#>       14) V1>=0.041 22 11 M (0.50000000 0.50000000)  
#>         28) V11< 0.26695 11  0 M (1.00000000 0.00000000) *
#>         29) V11>=0.26695 11  0 R (0.00000000 1.00000000) *
#>       15) V1< 0.041 55  4 R (0.07272727 0.92727273) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V51>=0.01545 65 16 M (0.75384615 0.24615385)  
#>      4) V41< 0.65815 58  9 M (0.84482759 0.15517241)  
#>        8) V12>=0.08385 49  2 M (0.95918367 0.04081633) *
#>        9) V12< 0.08385 9  2 R (0.22222222 0.77777778) *
#>      5) V41>=0.65815 7  0 R (0.00000000 1.00000000) *
#>    3) V51< 0.01545 74 19 R (0.25675676 0.74324324)  
#>      6) V22>=0.9026 7  0 M (1.00000000 0.00000000) *
#>      7) V22< 0.9026 67 12 R (0.17910448 0.82089552)  
#>       14) V49>=0.06035 13  4 M (0.69230769 0.30769231) *
#>       15) V49< 0.06035 54  3 R (0.05555556 0.94444444) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V11>=0.1707 89 21 M (0.76404494 0.23595506)  
#>      4) V49>=0.0227 80 12 M (0.85000000 0.15000000)  
#>        8) V15< 0.56165 58  2 M (0.96551724 0.03448276) *
#>        9) V15>=0.56165 22 10 M (0.54545455 0.45454545)  
#>         18) V11>=0.33625 12  0 M (1.00000000 0.00000000) *
#>         19) V11< 0.33625 10  0 R (0.00000000 1.00000000) *
#>      5) V49< 0.0227 9  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1707 50 10 R (0.20000000 0.80000000)  
#>      6) V32< 0.12165 9  1 M (0.88888889 0.11111111) *
#>      7) V32>=0.12165 41  2 R (0.04878049 0.95121951) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 50 M (0.64028777 0.35971223)  
#>    2) V52>=0.0107 80 13 M (0.83750000 0.16250000)  
#>      4) V38< 0.678 71  6 M (0.91549296 0.08450704)  
#>        8) V39>=0.1481 64  2 M (0.96875000 0.03125000) *
#>        9) V39< 0.1481 7  3 R (0.42857143 0.57142857) *
#>      5) V38>=0.678 9  2 R (0.22222222 0.77777778) *
#>    3) V52< 0.0107 59 22 R (0.37288136 0.62711864)  
#>      6) V11>=0.24655 22  5 M (0.77272727 0.22727273)  
#>       12) V16< 0.3302 15  0 M (1.00000000 0.00000000) *
#>       13) V16>=0.3302 7  2 R (0.28571429 0.71428571) *
#>      7) V11< 0.24655 37  5 R (0.13513514 0.86486486)  
#>       14) V27< 0.4949 8  3 M (0.62500000 0.37500000) *
#>       15) V27>=0.4949 29  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 M (0.60431655 0.39568345)  
#>    2) V45>=0.1953 49  7 M (0.85714286 0.14285714)  
#>      4) V48>=0.03825 42  1 M (0.97619048 0.02380952) *
#>      5) V48< 0.03825 7  1 R (0.14285714 0.85714286) *
#>    3) V45< 0.1953 90 42 R (0.46666667 0.53333333)  
#>      6) V34< 0.32975 36  7 M (0.80555556 0.19444444)  
#>       12) V11>=0.17095 29  1 M (0.96551724 0.03448276) *
#>       13) V11< 0.17095 7  1 R (0.14285714 0.85714286) *
#>      7) V34>=0.32975 54 13 R (0.24074074 0.75925926)  
#>       14) V28< 0.4473 12  3 M (0.75000000 0.25000000) *
#>       15) V28>=0.4473 42  4 R (0.09523810 0.90476190)  
#>         30) V5< 0.02555 7  3 M (0.57142857 0.42857143) *
#>         31) V5>=0.02555 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V21>=0.83575 45  4 M (0.91111111 0.08888889)  
#>      4) V30>=0.28515 38  0 M (1.00000000 0.00000000) *
#>      5) V30< 0.28515 7  3 R (0.42857143 0.57142857) *
#>    3) V21< 0.83575 94 40 R (0.42553191 0.57446809)  
#>      6) V28>=0.9055 16  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.9055 78 24 R (0.30769231 0.69230769)  
#>       14) V52>=0.02125 14  1 M (0.92857143 0.07142857) *
#>       15) V52< 0.02125 64 11 R (0.17187500 0.82812500)  
#>         30) V25>=0.91535 11  2 M (0.81818182 0.18181818) *
#>         31) V25< 0.91535 53  2 R (0.03773585 0.96226415) *
#> 
#> 
#> $weights
#>  [1] 0.6665923 1.0000614 0.7305476 0.7196378 0.7510207 0.6088237 0.7598963
#>  [8] 0.7619206 0.7148673 0.8713285
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0542387 5.5304576
#>   [2,] 2.4748250 5.1098713
#>   [3,] 2.3862915 5.1984048
#>   [4,] 1.3707444 6.2139519
#>   [5,] 0.0000000 7.5846963
#>   [6,] 1.3284616 6.2562347
#>   [7,] 1.6088851 5.9758112
#>   [8,] 2.3508627 5.2338336
#>   [9,] 3.0363813 4.5483150
#>  [10,] 2.1950196 5.3896767
#>  [11,] 0.8713285 6.7133678
#>  [12,] 2.9249491 4.6597472
#>  [13,] 3.2101431 4.3745531
#>  [14,] 2.4454763 5.1392200
#>  [15,] 2.1950196 5.3896767
#>  [16,] 1.4767879 6.1079083
#>  [17,] 1.4454149 6.1392813
#>  [18,] 2.2106999 5.3739964
#>  [19,] 2.2073356 5.3773607
#>  [20,] 1.6332491 5.9514471
#>  [21,] 2.4748250 5.1098713
#>  [22,] 3.6283030 3.9563933
#>  [23,] 0.0000000 7.5846963
#>  [24,] 2.1012920 5.4834043
#>  [25,] 0.0000000 7.5846963
#>  [26,] 0.7305476 6.8541487
#>  [27,] 3.2073969 4.3772993
#>  [28,] 2.1461264 5.4385699
#>  [29,] 1.4454149 6.1392813
#>  [30,] 0.8713285 6.7133678
#>  [31,] 1.4454149 6.1392813
#>  [32,] 0.7148673 6.8698290
#>  [33,] 1.6223492 5.9623471
#>  [34,] 0.7510207 6.8336756
#>  [35,] 0.0000000 7.5846963
#>  [36,] 0.8713285 6.7133678
#>  [37,] 0.0000000 7.5846963
#>  [38,] 1.0000614 6.5846349
#>  [39,] 0.8713285 6.7133678
#>  [40,] 0.7148673 6.8698290
#>  [41,] 0.7619206 6.8227756
#>  [42,] 0.0000000 7.5846963
#>  [43,] 0.7619206 6.8227756
#>  [44,] 1.4924683 6.0922280
#>  [45,] 0.0000000 7.5846963
#>  [46,] 0.0000000 7.5846963
#>  [47,] 1.5109170 6.0737793
#>  [48,] 1.8713899 5.7133064
#>  [49,] 2.5910277 4.9936686
#>  [50,] 1.7510820 5.8336142
#>  [51,] 2.9508108 4.6338855
#>  [52,] 3.1013534 4.4833429
#>  [53,] 2.3237524 5.2609439
#>  [54,] 2.4802136 5.1044827
#>  [55,] 1.4454149 6.1392813
#>  [56,] 2.0433289 5.5413674
#>  [57,] 2.0835874 5.5011089
#>  [58,] 2.1010974 5.4835989
#>  [59,] 1.4924683 6.0922280
#>  [60,] 1.5129413 6.0717550
#>  [61,] 2.8630180 4.7216782
#>  [62,] 0.7619206 6.8227756
#>  [63,] 5.0467141 2.5379821
#>  [64,] 6.1984662 1.3862301
#>  [65,] 6.3092803 1.2754160
#>  [66,] 6.2159762 1.3687201
#>  [67,] 3.9584176 3.6262787
#>  [68,] 5.4276601 2.1570362
#>  [69,] 4.6923197 2.8923766
#>  [70,] 5.1670220 2.4176743
#>  [71,] 6.8336756 0.7510207
#>  [72,] 6.0628793 1.5218170
#>  [73,] 7.5846963 0.0000000
#>  [74,] 6.8227756 0.7619206
#>  [75,] 6.9758725 0.6088237
#>  [76,] 7.5846963 0.0000000
#>  [77,] 6.8650584 0.7196378
#>  [78,] 6.1984662 1.3862301
#>  [79,] 5.4256357 2.1590605
#>  [80,] 6.9181040 0.6665923
#>  [81,] 5.4474455 2.1372508
#>  [82,] 6.1561834 1.4285129
#>  [83,] 6.8227756 0.7619206
#>  [84,] 6.1561834 1.4285129
#>  [85,] 5.9514471 1.6332491
#>  [86,] 5.4834043 2.1012920
#>  [87,] 6.2139519 1.3707444
#>  [88,] 6.7133678 0.8713285
#>  [89,] 6.8650584 0.7196378
#>  [90,] 6.7133678 0.8713285
#>  [91,] 7.5846963 0.0000000
#>  [92,] 6.2139519 1.3707444
#>  [93,] 4.6875492 2.8971471
#>  [94,] 6.9181040 0.6665923
#>  [95,] 6.7133678 0.8713285
#>  [96,] 6.0942523 1.4904439
#>  [97,] 6.8541487 0.7305476
#>  [98,] 5.9623471 1.6223492
#>  [99,] 4.7983633 2.7863330
#> [100,] 7.5846963 0.0000000
#> [101,] 6.1670833 1.4176130
#> [102,] 6.8336756 0.7510207
#> [103,] 5.4051627 2.1795336
#> [104,] 6.8336756 0.7510207
#> [105,] 6.1670833 1.4176130
#> [106,] 5.9623471 1.6223492
#> [107,] 5.4943042 2.0903920
#> [108,] 6.0467755 1.5379208
#> [109,] 6.1031280 1.4815683
#> [110,] 5.4385699 2.1461264
#> [111,] 4.6855249 2.8991714
#> [112,] 4.5099232 3.0747731
#> [113,] 6.2610052 1.3236910
#> [114,] 6.2610052 1.3236910
#> [115,] 5.3446477 2.2400486
#> [116,] 6.1561834 1.4285129
#> [117,] 6.2139519 1.3707444
#> [118,] 5.4943042 2.0903920
#> [119,] 5.5052140 2.0794822
#> [120,] 7.5846963 0.0000000
#> [121,] 7.5846963 0.0000000
#> [122,] 7.5846963 0.0000000
#> [123,] 6.0467755 1.5379208
#> [124,] 7.5846963 0.0000000
#> [125,] 7.5846963 0.0000000
#> [126,] 7.5846963 0.0000000
#> [127,] 6.8698290 0.7148673
#> [128,] 6.8541487 0.7305476
#> [129,] 7.5846963 0.0000000
#> [130,] 6.2453249 1.3393714
#> [131,] 6.8227756 0.7619206
#> [132,] 7.5846963 0.0000000
#> [133,] 6.0717550 1.5129413
#> [134,] 6.8227756 0.7619206
#> [135,] 7.5846963 0.0000000
#> [136,] 6.8541487 0.7305476
#> [137,] 7.5846963 0.0000000
#> [138,] 6.2453249 1.3393714
#> [139,] 6.0628793 1.5218170
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.27083994 0.72916006
#>   [2,] 0.32629190 0.67370810
#>   [3,] 0.31461925 0.68538075
#>   [4,] 0.18072502 0.81927498
#>   [5,] 0.00000000 1.00000000
#>   [6,] 0.17515027 0.82484973
#>   [7,] 0.21212255 0.78787745
#>   [8,] 0.30994816 0.69005184
#>   [9,] 0.40032998 0.59967002
#>  [10,] 0.28940111 0.71059889
#>  [11,] 0.11487982 0.88512018
#>  [12,] 0.38563826 0.61436174
#>  [13,] 0.42323951 0.57676049
#>  [14,] 0.32242244 0.67757756
#>  [15,] 0.28940111 0.71059889
#>  [16,] 0.19470627 0.80529373
#>  [17,] 0.19056992 0.80943008
#>  [18,] 0.29146848 0.70853152
#>  [19,] 0.29102491 0.70897509
#>  [20,] 0.21533481 0.78466519
#>  [21,] 0.32629190 0.67370810
#>  [22,] 0.47837156 0.52162844
#>  [23,] 0.00000000 1.00000000
#>  [24,] 0.27704366 0.72295634
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.09631864 0.90368136
#>  [27,] 0.42287744 0.57712256
#>  [28,] 0.28295483 0.71704517
#>  [29,] 0.19056992 0.80943008
#>  [30,] 0.11487982 0.88512018
#>  [31,] 0.19056992 0.80943008
#>  [32,] 0.09425128 0.90574872
#>  [33,] 0.21389771 0.78610229
#>  [34,] 0.09901790 0.90098210
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.11487982 0.88512018
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.13185253 0.86814747
#>  [39,] 0.11487982 0.88512018
#>  [40,] 0.09425128 0.90574872
#>  [41,] 0.10045500 0.89954500
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.10045500 0.89954500
#>  [44,] 0.19677364 0.80322636
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.19920600 0.80079400
#>  [48,] 0.24673234 0.75326766
#>  [49,] 0.34161259 0.65838741
#>  [50,] 0.23087042 0.76912958
#>  [51,] 0.38904798 0.61095202
#>  [52,] 0.40889618 0.59110382
#>  [53,] 0.30637383 0.69362617
#>  [54,] 0.32700236 0.67299764
#>  [55,] 0.19056992 0.80943008
#>  [56,] 0.26940154 0.73059846
#>  [57,] 0.27470940 0.72529060
#>  [58,] 0.27701800 0.72298200
#>  [59,] 0.19677364 0.80322636
#>  [60,] 0.19947289 0.80052711
#>  [61,] 0.37747300 0.62252700
#>  [62,] 0.10045500 0.89954500
#>  [63,] 0.66538118 0.33461882
#>  [64,] 0.81723328 0.18276672
#>  [65,] 0.83184350 0.16815650
#>  [66,] 0.81954188 0.18045812
#>  [67,] 0.52189534 0.47810466
#>  [68,] 0.71560678 0.28439322
#>  [69,] 0.61865624 0.38134376
#>  [70,] 0.68124309 0.31875691
#>  [71,] 0.90098210 0.09901790
#>  [72,] 0.79935690 0.20064310
#>  [73,] 1.00000000 0.00000000
#>  [74,] 0.89954500 0.10045500
#>  [75,] 0.91972998 0.08027002
#>  [76,] 1.00000000 0.00000000
#>  [77,] 0.90511976 0.09488024
#>  [78,] 0.81723328 0.18276672
#>  [79,] 0.71533988 0.28466012
#>  [80,] 0.91211352 0.08788648
#>  [81,] 0.71821538 0.28178462
#>  [82,] 0.81165852 0.18834148
#>  [83,] 0.89954500 0.10045500
#>  [84,] 0.81165852 0.18834148
#>  [85,] 0.78466519 0.21533481
#>  [86,] 0.72295634 0.27704366
#>  [87,] 0.81927498 0.18072502
#>  [88,] 0.88512018 0.11487982
#>  [89,] 0.90511976 0.09488024
#>  [90,] 0.88512018 0.11487982
#>  [91,] 1.00000000 0.00000000
#>  [92,] 0.81927498 0.18072502
#>  [93,] 0.61802728 0.38197272
#>  [94,] 0.91211352 0.08788648
#>  [95,] 0.88512018 0.11487982
#>  [96,] 0.80349326 0.19650674
#>  [97,] 0.90368136 0.09631864
#>  [98,] 0.78610229 0.21389771
#>  [99,] 0.63263750 0.36736250
#> [100,] 1.00000000 0.00000000
#> [101,] 0.81309562 0.18690438
#> [102,] 0.90098210 0.09901790
#> [103,] 0.71264062 0.28735938
#> [104,] 0.90098210 0.09901790
#> [105,] 0.81309562 0.18690438
#> [106,] 0.78610229 0.21389771
#> [107,] 0.72439344 0.27560656
#> [108,] 0.79723370 0.20276630
#> [109,] 0.80466346 0.19533654
#> [110,] 0.71704517 0.28295483
#> [111,] 0.61776038 0.38223962
#> [112,] 0.59460828 0.40539172
#> [113,] 0.82547870 0.17452130
#> [114,] 0.82547870 0.17452130
#> [115,] 0.70466206 0.29533794
#> [116,] 0.81165852 0.18834148
#> [117,] 0.81927498 0.18072502
#> [118,] 0.72439344 0.27560656
#> [119,] 0.72583184 0.27416816
#> [120,] 1.00000000 0.00000000
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 0.79723370 0.20276630
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 1.00000000 0.00000000
#> [127,] 0.90574872 0.09425128
#> [128,] 0.90368136 0.09631864
#> [129,] 1.00000000 0.00000000
#> [130,] 0.82341134 0.17658866
#> [131,] 0.89954500 0.10045500
#> [132,] 1.00000000 0.00000000
#> [133,] 0.80052711 0.19947289
#> [134,] 0.89954500 0.10045500
#> [135,] 1.00000000 0.00000000
#> [136,] 0.90368136 0.09631864
#> [137,] 1.00000000 0.00000000
#> [138,] 0.82341134 0.17658866
#> [139,] 0.79935690 0.20064310
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.0958179  6.2813589 24.6308222  3.2522327  0.0000000  0.0000000  1.0877392 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.8537226  4.2895858  1.5516463  0.0000000  0.0000000  0.0000000  3.1798683 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  5.3044178  0.0000000  0.0000000  2.4588820  0.0000000  0.9491559  4.2748201 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.8553748  0.0000000  2.0130885  0.0000000  2.5053757 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  3.0615771  1.7300188  0.0000000  1.4876743  0.7134863  0.0000000  0.6592305 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  1.3804540  0.0000000  0.0000000  1.1793315  3.7586623  0.0000000  2.6660338 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.5149409  3.5744554  0.6925444  0.0000000  2.6477218  9.3499602  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 77 62 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V52        V10        V22        V17        V28        V45 
#> 24.6308222  9.3499602  6.2813589  5.3044178  4.2895858  4.2748201  3.7586623 
#>        V49        V12        V21        V35        V47        V51        V34 
#>  3.5744554  3.2522327  3.1798683  3.0615771  2.6660338  2.6477218  2.5053757 
#>        V25        V32        V16        V36        V18        V48        V38 
#>  2.4588820  2.0130885  1.8537226  1.7300188  1.5516463  1.5149409  1.4876743 
#>        V41        V44         V1        V15        V27        V30        V39 
#>  1.3804540  1.1793315  1.0958179  1.0877392  0.9491559  0.8553748  0.7134863 
#>         V5        V40        V13        V14        V19         V2        V20 
#>  0.6925444  0.6592305  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V23        V24        V26        V29         V3        V31        V33 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V37         V4        V42        V43        V46        V50        V53 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478