Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V12>=0.20075 86 22 M (0.74418605 0.25581395)  
#>      4) V27>=0.7101 51  3 M (0.94117647 0.05882353) *
#>      5) V27< 0.7101 35 16 R (0.45714286 0.54285714)  
#>       10) V31< 0.37445 13  2 M (0.84615385 0.15384615) *
#>       11) V31>=0.37445 22  5 R (0.22727273 0.77272727)  
#>         22) V41>=0.4912 7  2 M (0.71428571 0.28571429) *
#>         23) V41< 0.4912 15  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.20075 53  6 R (0.11320755 0.88679245)  
#>      6) V20>=0.8291 9  3 M (0.66666667 0.33333333) *
#>      7) V20< 0.8291 44  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V11>=0.19645 82 21 M (0.74390244 0.25609756)  
#>      4) V16< 0.67795 58  5 M (0.91379310 0.08620690)  
#>        8) V33>=0.1491 51  1 M (0.98039216 0.01960784) *
#>        9) V33< 0.1491 7  3 R (0.42857143 0.57142857) *
#>      5) V16>=0.67795 24  8 R (0.33333333 0.66666667)  
#>       10) V12>=0.38075 9  1 M (0.88888889 0.11111111) *
#>       11) V12< 0.38075 15  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.19645 57 10 R (0.17543860 0.82456140)  
#>      6) V45>=0.15735 14  5 M (0.64285714 0.35714286) *
#>      7) V45< 0.15735 43  1 R (0.02325581 0.97674419) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 R (0.41007194 0.58992806)  
#>    2) V9>=0.10115 92 36 M (0.60869565 0.39130435)  
#>      4) V47>=0.0614 67 15 M (0.77611940 0.22388060)  
#>        8) V18< 0.6759 44  2 M (0.95454545 0.04545455) *
#>        9) V18>=0.6759 23 10 R (0.43478261 0.56521739)  
#>         18) V37< 0.49955 13  3 M (0.76923077 0.23076923) *
#>         19) V37>=0.49955 10  0 R (0.00000000 1.00000000) *
#>      5) V47< 0.0614 25  4 R (0.16000000 0.84000000)  
#>       10) V45>=0.1242 7  3 M (0.57142857 0.42857143) *
#>       11) V45< 0.1242 18  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.10115 47  1 R (0.02127660 0.97872340) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V28>=0.895 29  3 M (0.89655172 0.10344828) *
#>    3) V28< 0.895 110 32 R (0.29090909 0.70909091)  
#>      6) V52>=0.0211 12  0 M (1.00000000 0.00000000) *
#>      7) V52< 0.0211 98 20 R (0.20408163 0.79591837)  
#>       14) V2>=0.0781 10  2 M (0.80000000 0.20000000) *
#>       15) V2< 0.0781 88 12 R (0.13636364 0.86363636)  
#>         30) V19>=0.80275 22  9 R (0.40909091 0.59090909)  
#>           60) V36>=0.20565 11  2 M (0.81818182 0.18181818) *
#>           61) V36< 0.20565 11  0 R (0.00000000 1.00000000) *
#>         31) V19< 0.80275 66  3 R (0.04545455 0.95454545) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 R (0.38129496 0.61870504)  
#>    2) V44>=0.41255 18  1 M (0.94444444 0.05555556) *
#>    3) V44< 0.41255 121 36 R (0.29752066 0.70247934)  
#>      6) V49>=0.0447 48 22 M (0.54166667 0.45833333)  
#>       12) V36< 0.46475 26  5 M (0.80769231 0.19230769)  
#>         24) V23>=0.71065 18  0 M (1.00000000 0.00000000) *
#>         25) V23< 0.71065 8  3 R (0.37500000 0.62500000) *
#>       13) V36>=0.46475 22  5 R (0.22727273 0.77272727)  
#>         26) V48>=0.17035 8  3 M (0.62500000 0.37500000) *
#>         27) V48< 0.17035 14  0 R (0.00000000 1.00000000) *
#>      7) V49< 0.0447 73 10 R (0.13698630 0.86301370)  
#>       14) V16< 0.20195 24  9 R (0.37500000 0.62500000)  
#>         28) V11>=0.1795 9  0 M (1.00000000 0.00000000) *
#>         29) V11< 0.1795 15  0 R (0.00000000 1.00000000) *
#>       15) V16>=0.20195 49  1 R (0.02040816 0.97959184) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.43165468 0.56834532)  
#>    2) V48>=0.04885 94 39 M (0.58510638 0.41489362)  
#>      4) V37< 0.4189 57 10 M (0.82456140 0.17543860)  
#>        8) V51>=0.0134 36  0 M (1.00000000 0.00000000) *
#>        9) V51< 0.0134 21 10 M (0.52380952 0.47619048)  
#>         18) V43>=0.231 9  0 M (1.00000000 0.00000000) *
#>         19) V43< 0.231 12  2 R (0.16666667 0.83333333) *
#>      5) V37>=0.4189 37  8 R (0.21621622 0.78378378)  
#>       10) V17< 0.19045 10  3 M (0.70000000 0.30000000) *
#>       11) V17>=0.19045 27  1 R (0.03703704 0.96296296) *
#>    3) V48< 0.04885 45  5 R (0.11111111 0.88888889)  
#>      6) V54>=0.0136 9  4 M (0.55555556 0.44444444) *
#>      7) V54< 0.0136 36  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V55< 0.00635 65 15 M (0.76923077 0.23076923)  
#>      4) V47>=0.06295 58  8 M (0.86206897 0.13793103)  
#>        8) V12>=0.1421 50  3 M (0.94000000 0.06000000) *
#>        9) V12< 0.1421 8  3 R (0.37500000 0.62500000) *
#>      5) V47< 0.06295 7  0 R (0.00000000 1.00000000) *
#>    3) V55>=0.00635 74 22 R (0.29729730 0.70270270)  
#>      6) V54>=0.0225 12  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.0225 62 10 R (0.16129032 0.83870968)  
#>       14) V23>=0.808 8  2 M (0.75000000 0.25000000) *
#>       15) V23< 0.808 54  4 R (0.07407407 0.92592593)  
#>         30) V53>=0.0163 7  3 M (0.57142857 0.42857143) *
#>         31) V53< 0.0163 47  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 57 R (0.41007194 0.58992806)  
#>   2) V11>=0.20195 68 23 M (0.66176471 0.33823529)  
#>     4) V6< 0.1283 48  8 M (0.83333333 0.16666667)  
#>       8) V32>=0.30005 36  1 M (0.97222222 0.02777778) *
#>       9) V32< 0.30005 12  5 R (0.41666667 0.58333333) *
#>     5) V6>=0.1283 20  5 R (0.25000000 0.75000000) *
#>   3) V11< 0.20195 71 12 R (0.16901408 0.83098592)  
#>     6) V1>=0.04005 7  1 M (0.85714286 0.14285714) *
#>     7) V1< 0.04005 64  6 R (0.09375000 0.90625000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.5179856 0.4820144)  
#>    2) V12>=0.16505 89 24 M (0.7303371 0.2696629)  
#>      4) V40< 0.58555 80 16 M (0.8000000 0.2000000)  
#>        8) V47>=0.0614 73 10 M (0.8630137 0.1369863)  
#>         16) V10>=0.0627 64  4 M (0.9375000 0.0625000) *
#>         17) V10< 0.0627 9  3 R (0.3333333 0.6666667) *
#>        9) V47< 0.0614 7  1 R (0.1428571 0.8571429) *
#>      5) V40>=0.58555 9  1 R (0.1111111 0.8888889) *
#>    3) V12< 0.16505 50  7 R (0.1400000 0.8600000)  
#>      6) V47>=0.19755 10  3 M (0.7000000 0.3000000) *
#>      7) V47< 0.19755 40  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V11>=0.10985 111 45 M (0.59459459 0.40540541)  
#>      4) V42>=0.2604 50  6 M (0.88000000 0.12000000)  
#>        8) V37< 0.6035 43  1 M (0.97674419 0.02325581) *
#>        9) V37>=0.6035 7  2 R (0.28571429 0.71428571) *
#>      5) V42< 0.2604 61 22 R (0.36065574 0.63934426)  
#>       10) V5< 0.0646 27  9 M (0.66666667 0.33333333)  
#>         20) V3>=0.01145 19  2 M (0.89473684 0.10526316) *
#>         21) V3< 0.01145 8  1 R (0.12500000 0.87500000) *
#>       11) V5>=0.0646 34  4 R (0.11764706 0.88235294)  
#>         22) V34< 0.19825 7  3 M (0.57142857 0.42857143) *
#>         23) V34>=0.19825 27  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.10985 28  1 R (0.03571429 0.96428571) *
#> 
#> 
#> $weights
#>  [1] 0.8355657 0.7375983 0.8252605 0.7594690 0.8308369 0.7930623 0.6211124
#>  [8] 0.5487079 0.6945134 0.7082379
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.8126931 4.5416712
#>   [2,] 1.5306606 5.8237037
#>   [3,] 2.7540131 4.6003512
#>   [4,] 2.9140798 4.4402845
#>   [5,] 2.3553396 4.9990247
#>   [6,] 2.1511915 5.2031728
#>   [7,] 0.0000000 7.3543643
#>   [8,] 2.2394352 5.1149291
#>   [9,] 2.2497405 5.1046238
#>  [10,] 2.3662264 4.9881379
#>  [11,] 3.0210740 4.3332903
#>  [12,] 2.2573722 5.0969921
#>  [13,] 0.7594690 6.5948953
#>  [14,] 1.4677069 5.8866574
#>  [15,] 1.5731641 5.7812002
#>  [16,] 1.5628588 5.7915055
#>  [17,] 2.7429844 4.6113799
#>  [18,] 1.5684352 5.7859291
#>  [19,] 1.4970673 5.8572969
#>  [20,] 0.7375983 6.6167660
#>  [21,] 0.6211124 6.7332519
#>  [22,] 0.7082379 6.6461264
#>  [23,] 2.1403495 5.2140148
#>  [24,] 2.3368659 5.0174984
#>  [25,] 1.5438036 5.8105607
#>  [26,] 2.2497405 5.1046238
#>  [27,] 2.6238027 4.7305616
#>  [28,] 2.1622202 5.1921441
#>  [29,] 1.3805814 5.9737829
#>  [30,] 1.4677069 5.8866574
#>  [31,] 1.5306606 5.8237037
#>  [32,] 2.2629485 5.0914158
#>  [33,] 2.3032726 5.0510917
#>  [34,] 2.9025143 4.4518500
#>  [35,] 2.0669486 5.2874157
#>  [36,] 2.1517730 5.2025913
#>  [37,] 0.0000000 7.3543643
#>  [38,] 0.7594690 6.5948953
#>  [39,] 2.2929674 5.0613969
#>  [40,] 2.9106605 4.4437038
#>  [41,] 0.7082379 6.6461264
#>  [42,] 0.7594690 6.5948953
#>  [43,] 0.0000000 7.3543643
#>  [44,] 0.7594690 6.5948953
#>  [45,] 1.3805814 5.9737829
#>  [46,] 0.7930623 6.5613020
#>  [47,] 1.6286281 5.7257362
#>  [48,] 1.4458362 5.9085281
#>  [49,] 0.0000000 7.3543643
#>  [50,] 0.7375983 6.6167660
#>  [51,] 0.7594690 6.5948953
#>  [52,] 1.6608262 5.6935381
#>  [53,] 0.8252605 6.5291038
#>  [54,] 1.4539824 5.9003819
#>  [55,] 2.8993186 4.4550457
#>  [56,] 2.2280117 5.1263526
#>  [57,] 2.2280117 5.1263526
#>  [58,] 2.2114183 5.1429460
#>  [59,] 1.4463729 5.9079914
#>  [60,] 1.5334983 5.8208659
#>  [61,] 1.4463729 5.9079914
#>  [62,] 1.6183228 5.7360415
#>  [63,] 0.7594690 6.5948953
#>  [64,] 1.5197738 5.8345905
#>  [65,] 1.9945441 5.3598202
#>  [66,] 2.3553396 4.9990247
#>  [67,] 0.7930623 6.5613020
#>  [68,] 5.0935951 2.2607692
#>  [69,] 5.9748195 1.3795448
#>  [70,] 5.1941772 2.1601871
#>  [71,] 5.0037536 2.3506107
#>  [72,] 5.9748195 1.3795448
#>  [73,] 4.2820591 3.0723052
#>  [74,] 5.8290141 1.5253502
#>  [75,] 4.6216851 2.7326791
#>  [76,] 5.1495590 2.2048053
#>  [77,] 5.9748195 1.3795448
#>  [78,] 6.0680581 1.2863062
#>  [79,] 5.1495590 2.2048053
#>  [80,] 6.0461873 1.3081770
#>  [81,] 5.9748195 1.3795448
#>  [82,] 6.5613020 0.7930623
#>  [83,] 6.8056564 0.5487079
#>  [84,] 5.2372212 2.1171431
#>  [85,] 5.9748195 1.3795448
#>  [86,] 6.8056564 0.5487079
#>  [87,] 5.3735447 1.9808196
#>  [88,] 6.0974185 1.2569458
#>  [89,] 5.7304651 1.6238992
#>  [90,] 4.9797238 2.3746405
#>  [91,] 5.4029052 1.9514591
#>  [92,] 5.2618528 2.0925115
#>  [93,] 5.9700906 1.3842737
#>  [94,] 5.2803062 2.0740581
#>  [95,] 7.3543643 0.0000000
#>  [96,] 5.6879617 1.6664026
#>  [97,] 6.8056564 0.5487079
#>  [98,] 5.1817572 2.1726071
#>  [99,] 6.5948953 0.7594690
#> [100,] 5.2372212 2.1171431
#> [101,] 5.1073196 2.2470447
#> [102,] 5.8105607 1.5438036
#> [103,] 4.7817928 2.5725715
#> [104,] 5.3305007 2.0238636
#> [105,] 5.8530641 1.5013002
#> [106,] 5.8530641 1.5013002
#> [107,] 5.3305007 2.0238636
#> [108,] 5.7640584 1.5903059
#> [109,] 5.0668493 2.2875150
#> [110,] 4.5237177 2.8306466
#> [111,] 5.3043562 2.0500081
#> [112,] 6.5948953 0.7594690
#> [113,] 7.3543643 0.0000000
#> [114,] 6.5948953 0.7594690
#> [115,] 5.1073196 2.2470447
#> [116,] 6.5948953 0.7594690
#> [117,] 4.4441589 2.9102054
#> [118,] 6.7332519 0.6211124
#> [119,] 6.7332519 0.6211124
#> [120,] 6.8056564 0.5487079
#> [121,] 6.7332519 0.6211124
#> [122,] 6.0461873 1.3081770
#> [123,] 6.5235274 0.8308369
#> [124,] 6.5235274 0.8308369
#> [125,] 5.9748195 1.3795448
#> [126,] 5.9024150 1.4519493
#> [127,] 5.9024150 1.4519493
#> [128,] 7.3543643 0.0000000
#> [129,] 5.9024150 1.4519493
#> [130,] 6.8056564 0.5487079
#> [131,] 7.3543643 0.0000000
#> [132,] 6.6461264 0.7082379
#> [133,] 6.0125941 1.3417702
#> [134,] 6.6461264 0.7082379
#> [135,] 6.8056564 0.5487079
#> [136,] 5.8530641 1.5013002
#> [137,] 7.3543643 0.0000000
#> [138,] 5.2531250 2.1012393
#> [139,] 5.8018330 1.5525313
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.38245224 0.61754776
#>   [2,] 0.20812956 0.79187044
#>   [3,] 0.37447331 0.62552669
#>   [4,] 0.39623816 0.60376184
#>   [5,] 0.32026420 0.67973580
#>   [6,] 0.29250543 0.70749457
#>   [7,] 0.00000000 1.00000000
#>   [8,] 0.30450425 0.69549575
#>   [9,] 0.30590550 0.69409450
#>  [10,] 0.32174451 0.67825549
#>  [11,] 0.41078656 0.58921344
#>  [12,] 0.30694321 0.69305679
#>  [13,] 0.10326780 0.89673220
#>  [14,] 0.19956951 0.80043049
#>  [15,] 0.21390891 0.78609109
#>  [16,] 0.21250767 0.78749233
#>  [17,] 0.37297369 0.62702631
#>  [18,] 0.21326591 0.78673409
#>  [19,] 0.20356176 0.79643824
#>  [20,] 0.10029396 0.89970604
#>  [21,] 0.08445494 0.91554506
#>  [22,] 0.09630171 0.90369829
#>  [23,] 0.29103121 0.70896879
#>  [24,] 0.31775226 0.68224774
#>  [25,] 0.20991666 0.79008334
#>  [26,] 0.30590550 0.69409450
#>  [27,] 0.35676812 0.64323188
#>  [28,] 0.29400505 0.70599495
#>  [29,] 0.18772274 0.81227726
#>  [30,] 0.19956951 0.80043049
#>  [31,] 0.20812956 0.79187044
#>  [32,] 0.30770145 0.69229855
#>  [33,] 0.31318446 0.68681554
#>  [34,] 0.39466556 0.60533444
#>  [35,] 0.28105061 0.71894939
#>  [36,] 0.29258451 0.70741549
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.10326780 0.89673220
#>  [39,] 0.31178322 0.68821678
#>  [40,] 0.39577323 0.60422677
#>  [41,] 0.09630171 0.90369829
#>  [42,] 0.10326780 0.89673220
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.10326780 0.89673220
#>  [45,] 0.18772274 0.81227726
#>  [46,] 0.10783560 0.89216440
#>  [47,] 0.22145055 0.77854945
#>  [48,] 0.19659567 0.80340433
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.10029396 0.89970604
#>  [51,] 0.10326780 0.89673220
#>  [52,] 0.22582866 0.77417134
#>  [53,] 0.11221371 0.88778629
#>  [54,] 0.19770334 0.80229666
#>  [55,] 0.39423103 0.60576897
#>  [56,] 0.30295096 0.69704904
#>  [57,] 0.30295096 0.69704904
#>  [58,] 0.30069469 0.69930531
#>  [59,] 0.19666865 0.80333135
#>  [60,] 0.20851542 0.79148458
#>  [61,] 0.19666865 0.80333135
#>  [62,] 0.22004931 0.77995069
#>  [63,] 0.10326780 0.89673220
#>  [64,] 0.20664925 0.79335075
#>  [65,] 0.27120551 0.72879449
#>  [66,] 0.32026420 0.67973580
#>  [67,] 0.10783560 0.89216440
#>  [68,] 0.69259489 0.30740511
#>  [69,] 0.81241822 0.18758178
#>  [70,] 0.70627140 0.29372860
#>  [71,] 0.68037881 0.31962119
#>  [72,] 0.81241822 0.18758178
#>  [73,] 0.58224735 0.41775265
#>  [74,] 0.79259251 0.20740749
#>  [75,] 0.62842755 0.37157245
#>  [76,] 0.70020451 0.29979549
#>  [77,] 0.81241822 0.18758178
#>  [78,] 0.82509620 0.17490380
#>  [79,] 0.70020451 0.29979549
#>  [80,] 0.82212236 0.17787764
#>  [81,] 0.81241822 0.18758178
#>  [82,] 0.89216440 0.10783560
#>  [83,] 0.92539016 0.07460984
#>  [84,] 0.71212426 0.28787574
#>  [85,] 0.81241822 0.18758178
#>  [86,] 0.92539016 0.07460984
#>  [87,] 0.73066066 0.26933934
#>  [88,] 0.82908845 0.17091155
#>  [89,] 0.77919245 0.22080755
#>  [90,] 0.67711139 0.32288861
#>  [91,] 0.73465291 0.26534709
#>  [92,] 0.71547350 0.28452650
#>  [93,] 0.81177521 0.18822479
#>  [94,] 0.71798268 0.28201732
#>  [95,] 1.00000000 0.00000000
#>  [96,] 0.77341310 0.22658690
#>  [97,] 0.92539016 0.07460984
#>  [98,] 0.70458261 0.29541739
#>  [99,] 0.89673220 0.10326780
#> [100,] 0.71212426 0.28787574
#> [101,] 0.69446106 0.30553894
#> [102,] 0.79008334 0.20991666
#> [103,] 0.65019797 0.34980203
#> [104,] 0.72480781 0.27519219
#> [105,] 0.79586269 0.20413731
#> [106,] 0.79586269 0.20413731
#> [107,] 0.72480781 0.27519219
#> [108,] 0.78376025 0.21623975
#> [109,] 0.68895816 0.31104184
#> [110,] 0.61510656 0.38489344
#> [111,] 0.72125285 0.27874715
#> [112,] 0.89673220 0.10326780
#> [113,] 1.00000000 0.00000000
#> [114,] 0.89673220 0.10326780
#> [115,] 0.69446106 0.30553894
#> [116,] 0.89673220 0.10326780
#> [117,] 0.60428865 0.39571135
#> [118,] 0.91554506 0.08445494
#> [119,] 0.91554506 0.08445494
#> [120,] 0.92539016 0.07460984
#> [121,] 0.91554506 0.08445494
#> [122,] 0.82212236 0.17787764
#> [123,] 0.88702805 0.11297195
#> [124,] 0.88702805 0.11297195
#> [125,] 0.81241822 0.18758178
#> [126,] 0.80257311 0.19742689
#> [127,] 0.80257311 0.19742689
#> [128,] 1.00000000 0.00000000
#> [129,] 0.80257311 0.19742689
#> [130,] 0.92539016 0.07460984
#> [131,] 1.00000000 0.00000000
#> [132,] 0.90369829 0.09630171
#> [133,] 0.81755456 0.18244544
#> [134,] 0.90369829 0.09630171
#> [135,] 0.92539016 0.07460984
#> [136,] 0.79586269 0.20413731
#> [137,] 1.00000000 0.00000000
#> [138,] 0.71428676 0.28571324
#> [139,] 0.78889660 0.21110340
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.0543888  1.0452915 11.6335520 12.1773640  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  3.0840476  1.3292038  1.7597498  0.8659091  1.5695339  1.4499770  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  1.9725563  0.0000000  0.0000000  0.0000000  2.1233184  3.3408926 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  1.2344889  0.0000000  1.3665854  0.7964938  0.7224368  0.6717839 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  3.2042255  5.9473767  0.0000000  0.0000000  0.0000000  1.3933658 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  1.0632487  2.7429024  1.4801065  2.8471395  2.2725469  0.0000000  7.1121398 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.6969144  2.0590289  1.6788480  0.0000000  1.2464055  2.6878526  0.6457136 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.2164611  2.5015301  0.0000000  0.0000000  0.0000000  0.0000000  1.3774658 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  4.6291546 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12        V11        V47        V37         V9        V48        V28 
#> 12.1773640 11.6335520  7.1121398  5.9473767  4.6291546  3.6969144  3.3408926 
#>        V54        V36        V16        V44        V42        V52        V55 
#>  3.2164611  3.2042255  3.0840476  2.8471395  2.7429024  2.6878526  2.5015301 
#>        V45        V27        V49        V23        V18         V5         V2 
#>  2.2725469  2.1233184  2.0590289  1.9725563  1.7597498  1.6788480  1.5695339 
#>        V43        V20        V40         V6        V31        V17        V51 
#>  1.4801065  1.4499770  1.3933658  1.3774658  1.3665854  1.3292038  1.2464055 
#>         V3        V41         V1        V10        V19        V32        V33 
#>  1.2344889  1.0632487  1.0543888  1.0452915  0.8659091  0.7964938  0.7224368 
#>        V34        V53        V13        V14        V15        V21        V22 
#>  0.6717839  0.6457136  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V25        V26        V29        V30        V35        V38 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V39         V4        V46        V50        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1304348