Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.21655 85 21 M (0.75294118 0.24705882)  
#>      4) V44>=0.19055 42  2 M (0.95238095 0.04761905) *
#>      5) V44< 0.19055 43 19 M (0.55813953 0.44186047)  
#>       10) V16< 0.22155 12  0 M (1.00000000 0.00000000) *
#>       11) V16>=0.22155 31 12 R (0.38709677 0.61290323)  
#>         22) V2>=0.08305 9  0 M (1.00000000 0.00000000) *
#>         23) V2< 0.08305 22  3 R (0.13636364 0.86363636) *
#>    3) V11< 0.21655 54 10 R (0.18518519 0.81481481)  
#>      6) V46>=0.16885 16  6 M (0.62500000 0.37500000) *
#>      7) V46< 0.16885 38  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 M (0.54676259 0.45323741)  
#>   2) V12>=0.20075 88 20 M (0.77272727 0.22727273)  
#>     4) V24>=0.5548 69  8 M (0.88405797 0.11594203) *
#>     5) V24< 0.5548 19  7 R (0.36842105 0.63157895) *
#>   3) V12< 0.20075 51  8 R (0.15686275 0.84313725)  
#>     6) V21>=0.87685 13  6 M (0.53846154 0.46153846) *
#>     7) V21< 0.87685 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 59 M (0.5755396 0.4244604)  
#>   2) V12>=0.1555 99 23 M (0.7676768 0.2323232)  
#>     4) V11>=0.1707 82 12 M (0.8536585 0.1463415)  
#>       8) V51>=0.01215 64  2 M (0.9687500 0.0312500) *
#>       9) V51< 0.01215 18  8 R (0.4444444 0.5555556) *
#>     5) V11< 0.1707 17  6 R (0.3529412 0.6470588) *
#>   3) V12< 0.1555 40  4 R (0.1000000 0.9000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 54 M (0.61151079 0.38848921)  
#>    2) V21>=0.5198 89 18 M (0.79775281 0.20224719)  
#>      4) V43>=0.18465 55  1 M (0.98181818 0.01818182) *
#>      5) V43< 0.18465 34 17 M (0.50000000 0.50000000)  
#>       10) V1>=0.02045 18  3 M (0.83333333 0.16666667) *
#>       11) V1< 0.02045 16  2 R (0.12500000 0.87500000) *
#>    3) V21< 0.5198 50 14 R (0.28000000 0.72000000)  
#>      6) V28>=0.93605 11  1 M (0.90909091 0.09090909) *
#>      7) V28< 0.93605 39  4 R (0.10256410 0.89743590) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V10>=0.22795 56  8 M (0.85714286 0.14285714) *
#>    3) V10< 0.22795 83 30 R (0.36144578 0.63855422)  
#>      6) V28>=0.93025 9  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.93025 74 21 R (0.28378378 0.71621622)  
#>       14) V21>=0.81325 20  5 M (0.75000000 0.25000000)  
#>         28) V12>=0.16485 13  0 M (1.00000000 0.00000000) *
#>         29) V12< 0.16485 7  2 R (0.28571429 0.71428571) *
#>       15) V21< 0.81325 54  6 R (0.11111111 0.88888889)  
#>         30) V2>=0.0611 7  3 M (0.57142857 0.42857143) *
#>         31) V2< 0.0611 47  2 R (0.04255319 0.95744681) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.5395683 0.4604317)  
#>    2) V58>=0.00615 66  9 M (0.8636364 0.1363636) *
#>    3) V58< 0.00615 73 18 R (0.2465753 0.7534247)  
#>      6) V44>=0.3931 9  0 M (1.0000000 0.0000000) *
#>      7) V44< 0.3931 64  9 R (0.1406250 0.8593750)  
#>       14) V55< 0.007 21  9 R (0.4285714 0.5714286)  
#>         28) V10>=0.1268 11  2 M (0.8181818 0.1818182) *
#>         29) V10< 0.1268 10  0 R (0.0000000 1.0000000) *
#>       15) V55>=0.007 43  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 R (0.41007194 0.58992806)  
#>    2) V31< 0.4495 51 15 M (0.70588235 0.29411765)  
#>      4) V4>=0.03785 31  1 M (0.96774194 0.03225806) *
#>      5) V4< 0.03785 20  6 R (0.30000000 0.70000000)  
#>       10) V23>=0.777 8  2 M (0.75000000 0.25000000) *
#>       11) V23< 0.777 12  0 R (0.00000000 1.00000000) *
#>    3) V31>=0.4495 88 21 R (0.23863636 0.76136364)  
#>      6) V44>=0.29295 20  5 M (0.75000000 0.25000000)  
#>       12) V19< 0.4663 13  0 M (1.00000000 0.00000000) *
#>       13) V19>=0.4663 7  2 R (0.28571429 0.71428571) *
#>      7) V44< 0.29295 68  6 R (0.08823529 0.91176471) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V44>=0.29295 30  3 M (0.90000000 0.10000000) *
#>    3) V44< 0.29295 109 39 R (0.35779817 0.64220183)  
#>      6) V38< 0.20315 43 16 M (0.62790698 0.37209302)  
#>       12) V58>=0.00385 33  6 M (0.81818182 0.18181818)  
#>         24) V50< 0.0199 25  0 M (1.00000000 0.00000000) *
#>         25) V50>=0.0199 8  2 R (0.25000000 0.75000000) *
#>       13) V58< 0.00385 10  0 R (0.00000000 1.00000000) *
#>      7) V38>=0.20315 66 12 R (0.18181818 0.81818182)  
#>       14) V54>=0.0182 10  1 M (0.90000000 0.10000000) *
#>       15) V54< 0.0182 56  3 R (0.05357143 0.94642857) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V36< 0.40335 88 23 M (0.73863636 0.26136364)  
#>      4) V11>=0.1111 70  9 M (0.87142857 0.12857143)  
#>        8) V15< 0.645 63  4 M (0.93650794 0.06349206)  
#>         16) V32>=0.2693 56  0 M (1.00000000 0.00000000) *
#>         17) V32< 0.2693 7  3 R (0.42857143 0.57142857) *
#>        9) V15>=0.645 7  2 R (0.28571429 0.71428571) *
#>      5) V11< 0.1111 18  4 R (0.22222222 0.77777778) *
#>    3) V36>=0.40335 51  9 R (0.17647059 0.82352941)  
#>      6) V1>=0.0466 7  2 M (0.71428571 0.28571429) *
#>      7) V1< 0.0466 44  4 R (0.09090909 0.90909091) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V31< 0.467 60 13 M (0.78333333 0.21666667)  
#>      4) V20>=0.49015 50  3 M (0.94000000 0.06000000) *
#>      5) V20< 0.49015 10  0 R (0.00000000 1.00000000) *
#>    3) V31>=0.467 79 24 R (0.30379747 0.69620253)  
#>      6) V48>=0.07635 33 12 M (0.63636364 0.36363636)  
#>       12) V60< 0.00675 18  0 M (1.00000000 0.00000000) *
#>       13) V60>=0.00675 15  3 R (0.20000000 0.80000000) *
#>      7) V48< 0.07635 46  3 R (0.06521739 0.93478261) *
#> 
#> 
#> $weights
#>  [1] 0.7346456 0.6436199 0.5775177 0.7310908 0.7268045 0.6540811 0.6252011
#>  [8] 0.6482544 0.7251490 0.7495703
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0334264 4.7825081
#>   [2,] 2.8466221 3.9693123
#>   [3,] 2.0081011 4.8078334
#>   [4,] 2.7605621 4.0553724
#>   [5,] 1.3808857 5.4350488
#>   [6,] 2.1119765 4.7039579
#>   [7,] 2.8510855 3.9648489
#>   [8,] 0.0000000 6.8159345
#>   [9,] 2.0519058 4.7640287
#>  [10,] 2.0183912 4.7975432
#>  [11,] 2.7599434 4.0559911
#>  [12,] 1.9479421 4.8679924
#>  [13,] 1.3043222 5.5116123
#>  [14,] 2.1058622 4.7100722
#>  [15,] 0.6540811 6.1618533
#>  [16,] 3.3396479 3.4762866
#>  [17,] 1.3887268 5.4272077
#>  [18,] 2.0081011 4.8078334
#>  [19,] 1.3747714 5.4411631
#>  [20,] 2.1034145 4.7125199
#>  [21,] 2.1199947 4.6959398
#>  [22,] 2.0060867 4.8098477
#>  [23,] 0.6436199 6.1723146
#>  [24,] 1.3503501 5.4655843
#>  [25,] 2.7829966 4.0329378
#>  [26,] 2.0287919 4.7871426
#>  [27,] 2.6063702 4.2095643
#>  [28,] 1.3023355 5.5135989
#>  [29,] 1.3734034 5.4425310
#>  [30,] 0.0000000 6.8159345
#>  [31,] 1.3808857 5.4350488
#>  [32,] 0.7346456 6.0812888
#>  [33,] 2.0369812 4.7789533
#>  [34,] 2.1242810 4.6916535
#>  [35,] 2.0287919 4.7871426
#>  [36,] 0.0000000 6.8159345
#>  [37,] 1.3792302 5.4367043
#>  [38,] 0.7310908 6.0848436
#>  [39,] 2.1830444 4.6328900
#>  [40,] 0.6540811 6.1618533
#>  [41,] 0.6540811 6.1618533
#>  [42,] 0.0000000 6.8159345
#>  [43,] 0.7268045 6.0891299
#>  [44,] 0.7495703 6.0663642
#>  [45,] 1.4036514 5.4122830
#>  [46,] 2.0081011 4.8078334
#>  [47,] 0.0000000 6.8159345
#>  [48,] 0.0000000 6.8159345
#>  [49,] 1.2734555 5.5424790
#>  [50,] 0.0000000 6.8159345
#>  [51,] 0.7346456 6.0812888
#>  [52,] 2.6724724 4.1434621
#>  [53,] 2.0287919 4.7871426
#>  [54,] 2.7494821 4.0664524
#>  [55,] 2.6975123 4.1184221
#>  [56,] 1.3747107 5.4412237
#>  [57,] 1.3747107 5.4412237
#>  [58,] 2.0139279 4.8020066
#>  [59,] 1.2792822 5.5366522
#>  [60,] 2.1097045 4.7062299
#>  [61,] 2.1119765 4.7039579
#>  [62,] 2.0228500 4.7930844
#>  [63,] 0.6436199 6.1723146
#>  [64,] 1.3023355 5.5135989
#>  [65,] 0.0000000 6.8159345
#>  [66,] 4.7930844 2.0228500
#>  [67,] 5.5240602 1.2918742
#>  [68,] 4.6233934 2.1925410
#>  [69,] 4.2118969 2.6040375
#>  [70,] 4.2014356 2.6144988
#>  [71,] 4.8650134 1.9509211
#>  [72,] 5.5037711 1.3121633
#>  [73,] 4.1303677 2.6855667
#>  [74,] 6.0907854 0.7251490
#>  [75,] 6.0907854 0.7251490
#>  [76,] 5.4425310 1.3734034
#>  [77,] 4.7065174 2.1094170
#>  [78,] 5.4330344 1.3829000
#>  [79,] 5.4330344 1.3829000
#>  [80,] 5.4560877 1.3598467
#>  [81,] 6.0907854 0.7251490
#>  [82,] 6.0907854 0.7251490
#>  [83,] 4.8650134 1.9509211
#>  [84,] 6.0907854 0.7251490
#>  [85,] 6.0907854 0.7251490
#>  [86,] 4.9649613 1.8509731
#>  [87,] 6.8159345 0.0000000
#>  [88,] 6.1723146 0.6436199
#>  [89,] 6.1723146 0.6436199
#>  [90,] 5.4227443 1.3931902
#>  [91,] 5.3412151 1.4747193
#>  [92,] 5.4596425 1.3562919
#>  [93,] 5.3317185 1.4842159
#>  [94,] 5.4888465 1.3270880
#>  [95,] 5.4227443 1.3931902
#>  [96,] 5.3352733 1.4806611
#>  [97,] 4.7929087 2.0230257
#>  [98,] 4.2200256 2.5959089
#>  [99,] 4.7989112 2.0170233
#> [100,] 5.4471656 1.3687689
#> [101,] 4.7062299 2.1097045
#> [102,] 4.2427913 2.5731431
#> [103,] 4.7805214 2.0354130
#> [104,] 3.4266165 3.3893179
#> [105,] 6.0907854 0.7251490
#> [106,] 4.7160747 2.0998598
#> [107,] 5.5132677 1.3026667
#> [108,] 5.3596946 1.4562399
#> [109,] 5.3580391 1.4578954
#> [110,] 5.5901624 1.2257720
#> [111,] 4.8637061 1.9522284
#> [112,] 4.2444468 2.5714877
#> [113,] 5.5132677 1.3026667
#> [114,] 5.5132677 1.3026667
#> [115,] 5.5132677 1.3026667
#> [116,] 4.1492889 2.6666456
#> [117,] 6.8159345 0.0000000
#> [118,] 6.2384168 0.5775177
#> [119,] 6.8159345 0.0000000
#> [120,] 4.0677597 2.7481748
#> [121,] 4.8785701 1.9373644
#> [122,] 4.1289998 2.6869347
#> [123,] 6.1907334 0.6252011
#> [124,] 6.8159345 0.0000000
#> [125,] 6.8159345 0.0000000
#> [126,] 6.8159345 0.0000000
#> [127,] 4.8637061 1.9522284
#> [128,] 5.5471135 1.2688209
#> [129,] 6.1907334 0.6252011
#> [130,] 5.5366522 1.2792822
#> [131,] 6.8159345 0.0000000
#> [132,] 6.1676801 0.6482544
#> [133,] 6.0848436 0.7310908
#> [134,] 5.4425310 1.3734034
#> [135,] 4.7929087 2.0230257
#> [136,] 4.9649613 1.8509731
#> [137,] 5.5424790 1.2734555
#> [138,] 4.9649613 1.8509731
#> [139,] 4.7929087 2.0230257
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.29833420 0.70166580
#>   [2,] 0.41764224 0.58235776
#>   [3,] 0.29461860 0.70538140
#>   [4,] 0.40501594 0.59498406
#>   [5,] 0.20259667 0.79740333
#>   [6,] 0.30985869 0.69014131
#>   [7,] 0.41829709 0.58170291
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.30104541 0.69895459
#>  [10,] 0.29612832 0.70387168
#>  [11,] 0.40492516 0.59507484
#>  [12,] 0.28579237 0.71420763
#>  [13,] 0.19136366 0.80863634
#>  [14,] 0.30896163 0.69103837
#>  [15,] 0.09596353 0.90403647
#>  [16,] 0.48997652 0.51002348
#>  [17,] 0.20374709 0.79625291
#>  [18,] 0.29461860 0.70538140
#>  [19,] 0.20169962 0.79830038
#>  [20,] 0.30860252 0.69139748
#>  [21,] 0.31103507 0.68896493
#>  [22,] 0.29432307 0.70567693
#>  [23,] 0.09442871 0.90557129
#>  [24,] 0.19811665 0.80188335
#>  [25,] 0.40830743 0.59169257
#>  [26,] 0.29765425 0.70234575
#>  [27,] 0.38239367 0.61760633
#>  [28,] 0.19107219 0.80892781
#>  [29,] 0.20149892 0.79850108
#>  [30,] 0.00000000 1.00000000
#>  [31,] 0.20259667 0.79740333
#>  [32,] 0.10778355 0.89221645
#>  [33,] 0.29885574 0.70114426
#>  [34,] 0.31166394 0.68833606
#>  [35,] 0.29765425 0.70234575
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.20235379 0.79764621
#>  [38,] 0.10726201 0.89273799
#>  [39,] 0.32028542 0.67971458
#>  [40,] 0.09596353 0.90403647
#>  [41,] 0.09596353 0.90403647
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.10663314 0.89336686
#>  [44,] 0.10997322 0.89002678
#>  [45,] 0.20593676 0.79406324
#>  [46,] 0.29461860 0.70538140
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.18683505 0.81316495
#>  [50,] 0.00000000 1.00000000
#>  [51,] 0.10778355 0.89221645
#>  [52,] 0.39209185 0.60790815
#>  [53,] 0.29765425 0.70234575
#>  [54,] 0.40339033 0.59660967
#>  [55,] 0.39576559 0.60423441
#>  [56,] 0.20169072 0.79830928
#>  [57,] 0.20169072 0.79830928
#>  [58,] 0.29547348 0.70452652
#>  [59,] 0.18768992 0.81231008
#>  [60,] 0.30952535 0.69047465
#>  [61,] 0.30985869 0.69014131
#>  [62,] 0.29678250 0.70321750
#>  [63,] 0.09442871 0.90557129
#>  [64,] 0.19107219 0.80892781
#>  [65,] 0.00000000 1.00000000
#>  [66,] 0.70321750 0.29678250
#>  [67,] 0.81046264 0.18953736
#>  [68,] 0.67832129 0.32167871
#>  [69,] 0.61794856 0.38205144
#>  [70,] 0.61641374 0.38358626
#>  [71,] 0.71377056 0.28622944
#>  [72,] 0.80748593 0.19251407
#>  [73,] 0.60598701 0.39401299
#>  [74,] 0.89360974 0.10639026
#>  [75,] 0.89360974 0.10639026
#>  [76,] 0.79850108 0.20149892
#>  [77,] 0.69051683 0.30948317
#>  [78,] 0.79710779 0.20289221
#>  [79,] 0.79710779 0.20289221
#>  [80,] 0.80049005 0.19950995
#>  [81,] 0.89360974 0.10639026
#>  [82,] 0.89360974 0.10639026
#>  [83,] 0.71377056 0.28622944
#>  [84,] 0.89360974 0.10639026
#>  [85,] 0.89360974 0.10639026
#>  [86,] 0.72843443 0.27156557
#>  [87,] 1.00000000 0.00000000
#>  [88,] 0.90557129 0.09442871
#>  [89,] 0.90557129 0.09442871
#>  [90,] 0.79559807 0.20440193
#>  [91,] 0.78363652 0.21636348
#>  [92,] 0.80101159 0.19898841
#>  [93,] 0.78224322 0.21775678
#>  [94,] 0.80529626 0.19470374
#>  [95,] 0.79559807 0.20440193
#>  [96,] 0.78276476 0.21723524
#>  [97,] 0.70319173 0.29680827
#>  [98,] 0.61914116 0.38085884
#>  [99,] 0.70407238 0.29592762
#> [100,] 0.79918104 0.20081896
#> [101,] 0.69047465 0.30952535
#> [102,] 0.62248124 0.37751876
#> [103,] 0.70137432 0.29862568
#> [104,] 0.50273613 0.49726387
#> [105,] 0.89360974 0.10639026
#> [106,] 0.69191902 0.30808098
#> [107,] 0.80887922 0.19112078
#> [108,] 0.78634773 0.21365227
#> [109,] 0.78610484 0.21389516
#> [110,] 0.82016082 0.17983918
#> [111,] 0.71357876 0.28642124
#> [112,] 0.62272412 0.37727588
#> [113,] 0.80887922 0.19112078
#> [114,] 0.80887922 0.19112078
#> [115,] 0.80887922 0.19112078
#> [116,] 0.60876302 0.39123698
#> [117,] 1.00000000 0.00000000
#> [118,] 0.91526948 0.08473052
#> [119,] 1.00000000 0.00000000
#> [120,] 0.59680147 0.40319853
#> [121,] 0.71575953 0.28424047
#> [122,] 0.60578631 0.39421369
#> [123,] 0.90827361 0.09172639
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 1.00000000 0.00000000
#> [127,] 0.71357876 0.28642124
#> [128,] 0.81384490 0.18615510
#> [129,] 0.90827361 0.09172639
#> [130,] 0.81231008 0.18768992
#> [131,] 1.00000000 0.00000000
#> [132,] 0.90489134 0.09510866
#> [133,] 0.89273799 0.10726201
#> [134,] 0.79850108 0.20149892
#> [135,] 0.70319173 0.29680827
#> [136,] 0.72843443 0.27156557
#> [137,] 0.81316495 0.18683505
#> [138,] 0.72843443 0.27156557
#> [139,] 0.70319173 0.29680827
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  2.9664689  5.0985694  8.7809086 10.4274254  0.0000000  0.0000000  1.1935624 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.4727427  0.0000000  0.0000000  0.8952941  2.9229487  3.4046894  7.5499310 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  1.0412959  1.5726357  0.0000000  0.0000000  0.0000000  4.3624309 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  6.3444054  0.9088399  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  4.5644116  0.0000000  2.0718803  0.0000000  2.0904954  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  2.1997838  9.2235513  0.0000000  1.9931429  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  2.8982935  0.0000000  0.0000000  1.3632475  1.3758260  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  2.4308651  1.0456263  0.0000000  0.0000000  7.3795161  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  2.4212121  0.0000000  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12        V44        V11        V21        V58        V31        V10 
#> 10.4274254  9.2235513  8.7809086  7.5499310  7.3795161  6.3444054  5.0985694 
#>        V36        V28        V20         V1         V2        V48        V54 
#>  4.5644116  4.3624309  3.4046894  2.9664689  2.9229487  2.8982935  2.4308651 
#>        V60        V43         V4        V38        V46        V24        V16 
#>  2.4212121  2.1997838  2.0904954  2.0718803  1.9931429  1.5726357  1.4727427 
#>        V51        V50        V15        V55        V23        V32        V19 
#>  1.3758260  1.3632475  1.1935624  1.0456263  1.0412959  0.9088399  0.8952941 
#>        V13        V14        V17        V18        V22        V25        V26 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V27        V29         V3        V30        V33        V34        V35 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V37        V39        V40        V41        V42        V45        V47 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V49         V5        V52        V53        V56        V57        V59 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V6         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623