Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V13>=0.20655 94 26 M (0.72340426 0.27659574)  
#>      4) V52>=0.00885 60  4 M (0.93333333 0.06666667) *
#>      5) V52< 0.00885 34 12 R (0.35294118 0.64705882)  
#>       10) V23>=0.70225 14  2 M (0.85714286 0.14285714) *
#>       11) V23< 0.70225 20  0 R (0.00000000 1.00000000) *
#>    3) V13< 0.20655 45  5 R (0.11111111 0.88888889) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 R (0.49640288 0.50359712)  
#>   2) V11>=0.1779 78 18 M (0.76923077 0.23076923)  
#>     4) V39>=0.13005 68 10 M (0.85294118 0.14705882) *
#>     5) V39< 0.13005 10  2 R (0.20000000 0.80000000) *
#>   3) V11< 0.1779 61  9 R (0.14754098 0.85245902)  
#>     6) V46>=0.1639 9  1 M (0.88888889 0.11111111) *
#>     7) V46< 0.1639 52  1 R (0.01923077 0.98076923) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.4460432 0.5539568)  
#>    2) V11>=0.1779 81 29 M (0.6419753 0.3580247)  
#>      4) V51>=0.01235 50  7 M (0.8600000 0.1400000)  
#>        8) V21>=0.56055 37  0 M (1.0000000 0.0000000) *
#>        9) V21< 0.56055 13  6 R (0.4615385 0.5384615) *
#>      5) V51< 0.01235 31  9 R (0.2903226 0.7096774)  
#>       10) V5< 0.07295 13  4 M (0.6923077 0.3076923) *
#>       11) V5>=0.07295 18  0 R (0.0000000 1.0000000) *
#>    3) V11< 0.1779 58 10 R (0.1724138 0.8275862)  
#>      6) V1>=0.03895 8  1 M (0.8750000 0.1250000) *
#>      7) V1< 0.03895 50  3 R (0.0600000 0.9400000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V37< 0.3369 82 19 M (0.76829268 0.23170732)  
#>      4) V44>=0.1143 60  6 M (0.90000000 0.10000000) *
#>      5) V44< 0.1143 22  9 R (0.40909091 0.59090909)  
#>       10) V32>=0.4295 9  1 M (0.88888889 0.11111111) *
#>       11) V32< 0.4295 13  1 R (0.07692308 0.92307692) *
#>    3) V37>=0.3369 57 18 R (0.31578947 0.68421053)  
#>      6) V44>=0.33045 14  2 M (0.85714286 0.14285714) *
#>      7) V44< 0.33045 43  6 R (0.13953488 0.86046512)  
#>       14) V31< 0.3362 8  2 M (0.75000000 0.25000000) *
#>       15) V31>=0.3362 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V50>=0.0112 104 33 M (0.68269231 0.31730769)  
#>      4) V36< 0.4388 63  9 M (0.85714286 0.14285714)  
#>        8) V49>=0.02205 56  4 M (0.92857143 0.07142857) *
#>        9) V49< 0.02205 7  2 R (0.28571429 0.71428571) *
#>      5) V36>=0.4388 41 17 R (0.41463415 0.58536585)  
#>       10) V57>=0.0053 24  7 M (0.70833333 0.29166667)  
#>         20) V44>=0.27315 15  0 M (1.00000000 0.00000000) *
#>         21) V44< 0.27315 9  2 R (0.22222222 0.77777778) *
#>       11) V57< 0.0053 17  0 R (0.00000000 1.00000000) *
#>    3) V50< 0.0112 35  5 R (0.14285714 0.85714286)  
#>      6) V1>=0.029 9  4 M (0.55555556 0.44444444) *
#>      7) V1< 0.029 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V28>=0.9549 23  1 M (0.95652174 0.04347826) *
#>    3) V28< 0.9549 116 36 R (0.31034483 0.68965517)  
#>      6) V1>=0.0482 13  0 M (1.00000000 0.00000000) *
#>      7) V1< 0.0482 103 23 R (0.22330097 0.77669903)  
#>       14) V55< 0.00335 16  3 M (0.81250000 0.18750000) *
#>       15) V55>=0.00335 87 10 R (0.11494253 0.88505747)  
#>         30) V54>=0.01725 7  3 M (0.57142857 0.42857143) *
#>         31) V54< 0.01725 80  6 R (0.07500000 0.92500000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V37< 0.19385 46  8 M (0.82608696 0.17391304)  
#>      4) V11>=0.11445 38  2 M (0.94736842 0.05263158) *
#>      5) V11< 0.11445 8  2 R (0.25000000 0.75000000) *
#>    3) V37>=0.19385 93 30 R (0.32258065 0.67741935)  
#>      6) V42>=0.30645 33 11 M (0.66666667 0.33333333)  
#>       12) V7< 0.14195 23  3 M (0.86956522 0.13043478) *
#>       13) V7>=0.14195 10  2 R (0.20000000 0.80000000) *
#>      7) V42< 0.30645 60  8 R (0.13333333 0.86666667) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V44>=0.26765 43  4 M (0.90697674 0.09302326) *
#>    3) V44< 0.26765 96 41 R (0.42708333 0.57291667)  
#>      6) V12>=0.23825 53 18 M (0.66037736 0.33962264)  
#>       12) V57< 0.00565 26  0 M (1.00000000 0.00000000) *
#>       13) V57>=0.00565 27  9 R (0.33333333 0.66666667)  
#>         26) V23>=0.79935 9  0 M (1.00000000 0.00000000) *
#>         27) V23< 0.79935 18  0 R (0.00000000 1.00000000) *
#>      7) V12< 0.23825 43  6 R (0.13953488 0.86046512)  
#>       14) V21>=0.8088 8  2 M (0.75000000 0.25000000) *
#>       15) V21< 0.8088 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 R (0.3956835 0.6043165)  
#>    2) V48>=0.0757 66 26 M (0.6060606 0.3939394)  
#>      4) V16< 0.5678 49 11 M (0.7755102 0.2244898)  
#>        8) V11>=0.19475 31  0 M (1.0000000 0.0000000) *
#>        9) V11< 0.19475 18  7 R (0.3888889 0.6111111) *
#>      5) V16>=0.5678 17  2 R (0.1176471 0.8823529) *
#>    3) V48< 0.0757 73 15 R (0.2054795 0.7945205)  
#>      6) V59>=0.0071 23 11 M (0.5217391 0.4782609)  
#>       12) V26>=0.4741 14  2 M (0.8571429 0.1428571) *
#>       13) V26< 0.4741 9  0 R (0.0000000 1.0000000) *
#>      7) V59< 0.0071 50  3 R (0.0600000 0.9400000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V12>=0.19625 75 12 M (0.84000000 0.16000000)  
#>     4) V52>=0.00465 64  4 M (0.93750000 0.06250000) *
#>     5) V52< 0.00465 11  3 R (0.27272727 0.72727273) *
#>   3) V12< 0.19625 64 18 R (0.28125000 0.71875000)  
#>     6) V4>=0.0543 17  3 M (0.82352941 0.17647059) *
#>     7) V4< 0.0543 47  4 R (0.08510638 0.91489362) *
#> 
#> 
#> $weights
#>  [1] 0.6665923 0.7046864 0.7271666 0.6937631 0.7072963 0.7378321 0.6209528
#>  [8] 1.0227016 0.6129014 0.6510750
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.4237610 4.7212065
#>   [2,] 2.0630577 5.0819098
#>   [3,] 2.0223537 5.1226139
#>   [4,] 2.1498148 4.9951527
#>   [5,] 2.1696851 4.9752825
#>   [6,] 0.7378321 6.4071354
#>   [7,] 1.7164647 5.4285028
#>   [8,] 2.7517930 4.3931746
#>   [9,] 1.4119827 5.7329848
#>  [10,] 1.9305687 5.2143988
#>  [11,] 2.6323832 4.5125843
#>  [12,] 0.0000000 7.1449675
#>  [13,] 3.0450552 4.0999123
#>  [14,] 2.3939802 4.7509873
#>  [15,] 1.7273880 5.4175796
#>  [16,] 2.0114304 5.1335371
#>  [17,] 2.6958294 4.4491382
#>  [18,] 2.0418825 5.1030851
#>  [19,] 1.9385406 5.2064270
#>  [20,] 2.4271259 4.7178416
#>  [21,] 2.0720047 5.0729628
#>  [22,] 1.3583713 5.7865962
#>  [23,] 2.6759591 4.4690084
#>  [24,] 0.6129014 6.5320661
#>  [25,] 1.4010595 5.7439081
#>  [26,] 1.3147159 5.8302516
#>  [27,] 1.9716863 5.1732813
#>  [28,] 1.9732568 5.1717107
#>  [29,] 0.0000000 7.1449675
#>  [30,] 0.7046864 6.4402811
#>  [31,] 2.0223537 5.1226139
#>  [32,] 2.4211511 4.7238164
#>  [33,] 3.6931789 3.4517887
#>  [34,] 0.6665923 6.4783753
#>  [35,] 0.0000000 7.1449675
#>  [36,] 0.6937631 6.4512044
#>  [37,] 1.4010595 5.7439081
#>  [38,] 2.1388916 5.0060760
#>  [39,] 0.6129014 6.5320661
#>  [40,] 0.0000000 7.1449675
#>  [41,] 0.7378321 6.4071354
#>  [42,] 0.6209528 6.5240148
#>  [43,] 0.6209528 6.5240148
#>  [44,] 0.6209528 6.5240148
#>  [45,] 1.2720278 5.8729397
#>  [46,] 1.3282491 5.8167184
#>  [47,] 0.6937631 6.4512044
#>  [48,] 2.1362817 5.0086859
#>  [49,] 1.0227016 6.1222660
#>  [50,] 2.3293661 4.8156014
#>  [51,] 2.4542968 4.6906707
#>  [52,] 2.4237610 4.7212065
#>  [53,] 1.6892938 5.4556737
#>  [54,] 2.7296500 4.4153175
#>  [55,] 1.4344629 5.7105047
#>  [56,] 2.4237610 4.7212065
#>  [57,] 2.0329355 5.1120320
#>  [58,] 2.0554994 5.0894681
#>  [59,] 1.3889071 5.7560604
#>  [60,] 1.3448382 5.8001294
#>  [61,] 1.3256392 5.8193284
#>  [62,] 1.3448382 5.8001294
#>  [63,] 2.0114304 5.1335371
#>  [64,] 2.0962035 5.0487641
#>  [65,] 0.7072963 6.4376712
#>  [66,] 4.5507580 2.5942096
#>  [67,] 4.7178416 2.4271259
#>  [68,] 5.5093646 1.6356030
#>  [69,] 5.7133723 1.4315953
#>  [70,] 5.8383030 1.3066645
#>  [71,] 4.5372248 2.6077428
#>  [72,] 5.7942340 1.3507335
#>  [73,] 5.7892061 1.3557614
#>  [74,] 5.1383073 2.0066602
#>  [75,] 5.7131146 1.4318529
#>  [76,] 5.1200834 2.0248841
#>  [77,] 5.9111134 1.2338542
#>  [78,] 5.1351076 2.0098599
#>  [79,] 7.1449675 0.0000000
#>  [80,] 6.5320661 0.6129014
#>  [81,] 5.8729397 1.2720278
#>  [82,] 5.7861827 1.3587849
#>  [83,] 5.7439081 1.4010595
#>  [84,] 6.5320661 0.6129014
#>  [85,] 5.7942340 1.3507335
#>  [86,] 7.1449675 0.0000000
#>  [87,] 7.1449675 0.0000000
#>  [88,] 4.4795181 2.6654494
#>  [89,] 4.4990512 2.6459164
#>  [90,] 6.4783753 0.6665923
#>  [91,] 5.2173502 1.9276173
#>  [92,] 5.8167184 1.3282491
#>  [93,] 5.8574225 1.2875450
#>  [94,] 5.2173502 1.9276173
#>  [95,] 5.2445211 1.9004464
#>  [96,] 6.4178010 0.7271666
#>  [97,] 4.4461147 2.6988528
#>  [98,] 5.0167415 2.1282260
#>  [99,] 6.4178010 0.7271666
#> [100,] 6.4178010 0.7271666
#> [101,] 6.1222660 1.0227016
#> [102,] 5.1717107 1.9732568
#> [103,] 4.4685154 2.6764522
#> [104,] 5.1195904 2.0253771
#> [105,] 4.4336209 2.7113466
#> [106,] 5.0976033 2.0473643
#> [107,] 5.7968482 1.3481193
#> [108,] 5.6799689 1.4649987
#> [109,] 5.7105047 1.4344629
#> [110,] 5.6799689 1.4649987
#> [111,] 5.7512087 1.3937588
#> [112,] 5.7861827 1.3587849
#> [113,] 5.6799689 1.4649987
#> [114,] 5.1431590 2.0018085
#> [115,] 3.8662795 3.2786880
#> [116,] 5.2038170 1.9411505
#> [117,] 5.8247698 1.3201977
#> [118,] 5.8247698 1.3201977
#> [119,] 5.7240378 1.4209297
#> [120,] 5.1351076 2.0098599
#> [121,] 6.5240148 0.6209528
#> [122,] 5.9111134 1.2338542
#> [123,] 5.0924195 2.0525480
#> [124,] 6.1222660 1.0227016
#> [125,] 6.4071354 0.7378321
#> [126,] 7.1449675 0.0000000
#> [127,] 6.4178010 0.7271666
#> [128,] 5.7105047 1.4344629
#> [129,] 6.4178010 0.7271666
#> [130,] 6.4178010 0.7271666
#> [131,] 6.4178010 0.7271666
#> [132,] 6.1222660 1.0227016
#> [133,] 6.4402811 0.7046864
#> [134,] 6.4402811 0.7046864
#> [135,] 5.7024490 1.4425185
#> [136,] 6.4402811 0.7046864
#> [137,] 6.4178010 0.7271666
#> [138,] 6.4178010 0.7271666
#> [139,] 5.7405432 1.4044244
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.33922632 0.66077368
#>   [2,] 0.28874277 0.71125723
#>   [3,] 0.28304589 0.71695411
#>   [4,] 0.30088518 0.69911482
#>   [5,] 0.30366619 0.69633381
#>   [6,] 0.10326599 0.89673401
#>   [7,] 0.24023408 0.75976592
#>   [8,] 0.38513723 0.61486277
#>   [9,] 0.19761919 0.80238081
#>  [10,] 0.27019978 0.72980022
#>  [11,] 0.36842479 0.63157521
#>  [12,] 0.00000000 1.00000000
#>  [13,] 0.42618182 0.57381818
#>  [14,] 0.33505824 0.66494176
#>  [15,] 0.24176288 0.75823712
#>  [16,] 0.28151708 0.71848292
#>  [17,] 0.37730464 0.62269536
#>  [18,] 0.28577911 0.71422089
#>  [19,] 0.27131552 0.72868448
#>  [20,] 0.33969727 0.66030273
#>  [21,] 0.28999498 0.71000502
#>  [22,] 0.19011582 0.80988418
#>  [23,] 0.37452362 0.62547638
#>  [24,] 0.08578085 0.91421915
#>  [25,] 0.19609039 0.80390961
#>  [26,] 0.18400586 0.81599414
#>  [27,] 0.27595455 0.72404545
#>  [28,] 0.27617436 0.72382564
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.09862696 0.90137304
#>  [31,] 0.28304589 0.71695411
#>  [32,] 0.33886103 0.66113897
#>  [33,] 0.51689233 0.48310767
#>  [34,] 0.09329535 0.90670465
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.09709815 0.90290185
#>  [37,] 0.19609039 0.80390961
#>  [38,] 0.29935637 0.70064363
#>  [39,] 0.08578085 0.91421915
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.10326599 0.89673401
#>  [42,] 0.08690771 0.91309229
#>  [43,] 0.08690771 0.91309229
#>  [44,] 0.08690771 0.91309229
#>  [45,] 0.17803129 0.82196871
#>  [46,] 0.18589995 0.81410005
#>  [47,] 0.09709815 0.90290185
#>  [48,] 0.29899109 0.70100891
#>  [49,] 0.14313593 0.85686407
#>  [50,] 0.32601493 0.67398507
#>  [51,] 0.34350006 0.65649994
#>  [52,] 0.33922632 0.66077368
#>  [53,] 0.23643128 0.76356872
#>  [54,] 0.38203813 0.61796187
#>  [55,] 0.20076549 0.79923451
#>  [56,] 0.33922632 0.66077368
#>  [57,] 0.28452690 0.71547310
#>  [58,] 0.28768492 0.71231508
#>  [59,] 0.19438956 0.80561044
#>  [60,] 0.18822173 0.81177827
#>  [61,] 0.18553467 0.81446533
#>  [62,] 0.18822173 0.81177827
#>  [63,] 0.28151708 0.71848292
#>  [64,] 0.29338180 0.70661820
#>  [65,] 0.09899224 0.90100776
#>  [66,] 0.63691793 0.36308207
#>  [67,] 0.66030273 0.33969727
#>  [68,] 0.77108322 0.22891678
#>  [69,] 0.79963586 0.20036414
#>  [70,] 0.81712100 0.18287900
#>  [71,] 0.63502385 0.36497615
#>  [72,] 0.81095316 0.18904684
#>  [73,] 0.81024946 0.18975054
#>  [74,] 0.71915055 0.28084945
#>  [75,] 0.79959980 0.20040020
#>  [76,] 0.71659995 0.28340005
#>  [77,] 0.82731144 0.17268856
#>  [78,] 0.71870272 0.28129728
#>  [79,] 1.00000000 0.00000000
#>  [80,] 0.91421915 0.08578085
#>  [81,] 0.82196871 0.17803129
#>  [82,] 0.80982630 0.19017370
#>  [83,] 0.80390961 0.19609039
#>  [84,] 0.91421915 0.08578085
#>  [85,] 0.81095316 0.18904684
#>  [86,] 1.00000000 0.00000000
#>  [87,] 1.00000000 0.00000000
#>  [88,] 0.62694730 0.37305270
#>  [89,] 0.62968112 0.37031888
#>  [90,] 0.90670465 0.09329535
#>  [91,] 0.73021329 0.26978671
#>  [92,] 0.81410005 0.18589995
#>  [93,] 0.81979694 0.18020306
#>  [94,] 0.73021329 0.26978671
#>  [95,] 0.73401608 0.26598392
#>  [96,] 0.89822675 0.10177325
#>  [97,] 0.62227221 0.37772779
#>  [98,] 0.70213636 0.29786364
#>  [99,] 0.89822675 0.10177325
#> [100,] 0.89822675 0.10177325
#> [101,] 0.85686407 0.14313593
#> [102,] 0.72382564 0.27617436
#> [103,] 0.62540737 0.37459263
#> [104,] 0.71653095 0.28346905
#> [105,] 0.62052359 0.37947641
#> [106,] 0.71345366 0.28654634
#> [107,] 0.81131904 0.18868096
#> [108,] 0.79496077 0.20503923
#> [109,] 0.79923451 0.20076549
#> [110,] 0.79496077 0.20503923
#> [111,] 0.80493140 0.19506860
#> [112,] 0.80982630 0.19017370
#> [113,] 0.79496077 0.20503923
#> [114,] 0.71982958 0.28017042
#> [115,] 0.54111926 0.45888074
#> [116,] 0.72831920 0.27168080
#> [117,] 0.81522691 0.18477309
#> [118,] 0.81522691 0.18477309
#> [119,] 0.80112860 0.19887140
#> [120,] 0.71870272 0.28129728
#> [121,] 0.91309229 0.08690771
#> [122,] 0.82731144 0.17268856
#> [123,] 0.71272815 0.28727185
#> [124,] 0.85686407 0.14313593
#> [125,] 0.89673401 0.10326599
#> [126,] 1.00000000 0.00000000
#> [127,] 0.89822675 0.10177325
#> [128,] 0.79923451 0.20076549
#> [129,] 0.89822675 0.10177325
#> [130,] 0.89822675 0.10177325
#> [131,] 0.89822675 0.10177325
#> [132,] 0.85686407 0.14313593
#> [133,] 0.90137304 0.09862696
#> [134,] 0.90137304 0.09862696
#> [135,] 0.79810706 0.20189294
#> [136,] 0.90137304 0.09862696
#> [137,] 0.89822675 0.10177325
#> [138,] 0.89822675 0.10177325
#> [139,] 0.80343866 0.19656134
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#>  5.776538  0.000000 11.254779  7.915894  4.424771  0.000000  0.000000  1.947821 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  3.359513  0.000000  5.916711 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#>  0.000000  0.000000  1.435245  0.000000  3.440469  0.000000  0.000000  0.000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#>  1.478460  1.415259  0.000000  0.000000  0.000000  2.001347  5.598073  0.000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#>  1.523858  2.578662  0.000000  0.000000  2.187885  0.000000  9.231063  0.000000 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#>  2.379004  0.000000  1.983398  1.058186  1.530638  3.140483  2.627391  4.406611 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#>  0.000000  0.680967  2.822967  0.000000  5.557444  0.000000  1.197692  0.000000 
#>       V60        V7        V8        V9 
#>  0.000000  1.128871  0.000000  0.000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V11       V44       V12       V23        V1       V37       V57       V13 
#> 11.254779  9.231063  7.915894  5.916711  5.776538  5.598073  5.557444  4.424771 
#>       V52       V28       V21       V50       V55       V51        V4       V46 
#>  4.406611  3.440469  3.359513  3.140483  2.822967  2.627391  2.578662  2.379004 
#>       V42       V36       V48       V16        V5       V39       V31       V26 
#>  2.187885  2.001347  1.983398  1.947821  1.530638  1.523858  1.478460  1.435245 
#>       V32       V59        V7       V49       V54       V10       V14       V15 
#>  1.415259  1.197692  1.128871  1.058186  0.680967  0.000000  0.000000  0.000000 
#>       V17       V18       V19        V2       V20       V22       V24       V25 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V27       V29        V3       V30       V33       V34       V35       V38 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V40       V41       V43       V45       V47       V53       V56       V58 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>        V6       V60        V8        V9 
#>  0.000000  0.000000  0.000000  0.000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478