Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.5179856 0.4820144)  
#>    2) V9>=0.13175 83 22 M (0.7349398 0.2650602)  
#>      4) V15< 0.5107 62  7 M (0.8870968 0.1129032) *
#>      5) V15>=0.5107 21  6 R (0.2857143 0.7142857)  
#>       10) V54>=0.0127 7  1 M (0.8571429 0.1428571) *
#>       11) V54< 0.0127 14  0 R (0.0000000 1.0000000) *
#>    3) V9< 0.13175 56 11 R (0.1964286 0.8035714)  
#>      6) V28>=0.94385 9  3 M (0.6666667 0.3333333) *
#>      7) V28< 0.94385 47  5 R (0.1063830 0.8936170) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V48>=0.0847 61 17 M (0.72131148 0.27868852)  
#>      4) V27>=0.71555 31  0 M (1.00000000 0.00000000) *
#>      5) V27< 0.71555 30 13 R (0.43333333 0.56666667)  
#>       10) V37< 0.1923 8  0 M (1.00000000 0.00000000) *
#>       11) V37>=0.1923 22  5 R (0.22727273 0.77272727)  
#>         22) V1>=0.041 7  2 M (0.71428571 0.28571429) *
#>         23) V1< 0.041 15  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.0847 78 19 R (0.24358974 0.75641026)  
#>      6) V13>=0.24965 38 19 M (0.50000000 0.50000000)  
#>       12) V23>=0.79225 14  1 M (0.92857143 0.07142857) *
#>       13) V23< 0.79225 24  6 R (0.25000000 0.75000000)  
#>         26) V52>=0.0112 8  2 M (0.75000000 0.25000000) *
#>         27) V52< 0.0112 16  0 R (0.00000000 1.00000000) *
#>      7) V13< 0.24965 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 64 R (0.46043165 0.53956835)  
#>   2) V31< 0.4513 64 15 M (0.76562500 0.23437500)  
#>     4) V38>=0.22785 48  5 M (0.89583333 0.10416667) *
#>     5) V38< 0.22785 16  6 R (0.37500000 0.62500000) *
#>   3) V31>=0.4513 75 15 R (0.20000000 0.80000000)  
#>     6) V48>=0.108 14  3 M (0.78571429 0.21428571) *
#>     7) V48< 0.108 61  4 R (0.06557377 0.93442623) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V11>=0.1998 81 25 M (0.69135802 0.30864198)  
#>      4) V47>=0.06225 67 12 M (0.82089552 0.17910448)  
#>        8) V39>=0.16175 57  5 M (0.91228070 0.08771930)  
#>         16) V29>=0.327 50  1 M (0.98000000 0.02000000) *
#>         17) V29< 0.327 7  3 R (0.42857143 0.57142857) *
#>        9) V39< 0.16175 10  3 R (0.30000000 0.70000000) *
#>      5) V47< 0.06225 14  1 R (0.07142857 0.92857143) *
#>    3) V11< 0.1998 58  9 R (0.15517241 0.84482759)  
#>      6) V7>=0.14045 12  5 M (0.58333333 0.41666667) *
#>      7) V7< 0.14045 46  2 R (0.04347826 0.95652174) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V11>=0.16095 97 29 M (0.70103093 0.29896907)  
#>      4) V47>=0.0356 86 18 M (0.79069767 0.20930233)  
#>        8) V51>=0.0129 55  1 M (0.98181818 0.01818182) *
#>        9) V51< 0.0129 31 14 R (0.45161290 0.54838710)  
#>         18) V8< 0.06885 9  0 M (1.00000000 0.00000000) *
#>         19) V8>=0.06885 22  5 R (0.22727273 0.77272727)  
#>           38) V12>=0.26835 9  4 M (0.55555556 0.44444444) *
#>           39) V12< 0.26835 13  0 R (0.00000000 1.00000000) *
#>      5) V47< 0.0356 11  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.16095 42  8 R (0.19047619 0.80952381)  
#>      6) V4>=0.05095 10  2 M (0.80000000 0.20000000) *
#>      7) V4< 0.05095 32  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V36< 0.505 89 26 M (0.70786517 0.29213483)  
#>      4) V51>=0.01275 60  7 M (0.88333333 0.11666667)  
#>        8) V2>=0.02175 52  1 M (0.98076923 0.01923077) *
#>        9) V2< 0.02175 8  2 R (0.25000000 0.75000000) *
#>      5) V51< 0.01275 29 10 R (0.34482759 0.65517241)  
#>       10) V43>=0.2269 8  1 M (0.87500000 0.12500000) *
#>       11) V43< 0.2269 21  3 R (0.14285714 0.85714286) *
#>    3) V36>=0.505 50  7 R (0.14000000 0.86000000)  
#>      6) V45>=0.2797 10  3 M (0.70000000 0.30000000) *
#>      7) V45< 0.2797 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V12>=0.1519 98 43 M (0.56122449 0.43877551)  
#>      4) V16< 0.687 79 25 M (0.68354430 0.31645570)  
#>        8) V36< 0.4926 56  9 M (0.83928571 0.16071429)  
#>         16) V51>=0.0075 48  3 M (0.93750000 0.06250000) *
#>         17) V51< 0.0075 8  2 R (0.25000000 0.75000000) *
#>        9) V36>=0.4926 23  7 R (0.30434783 0.69565217)  
#>         18) V4>=0.0606 8  1 M (0.87500000 0.12500000) *
#>         19) V4< 0.0606 15  0 R (0.00000000 1.00000000) *
#>      5) V16>=0.687 19  1 R (0.05263158 0.94736842) *
#>    3) V12< 0.1519 41  3 R (0.07317073 0.92682927) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V28>=0.958 26  0 M (1.00000000 0.00000000) *
#>    3) V28< 0.958 113 47 R (0.41592920 0.58407080)  
#>      6) V4>=0.05195 51 15 M (0.70588235 0.29411765)  
#>       12) V50>=0.01525 35  3 M (0.91428571 0.08571429) *
#>       13) V50< 0.01525 16  4 R (0.25000000 0.75000000) *
#>      7) V4< 0.05195 62 11 R (0.17741935 0.82258065)  
#>       14) V23>=0.76 14  4 M (0.71428571 0.28571429) *
#>       15) V23< 0.76 48  1 R (0.02083333 0.97916667) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V48>=0.05195 99 38 M (0.61616162 0.38383838)  
#>      4) V36< 0.4926 64 12 M (0.81250000 0.18750000)  
#>        8) V23>=0.71065 38  1 M (0.97368421 0.02631579) *
#>        9) V23< 0.71065 26 11 M (0.57692308 0.42307692)  
#>         18) V20< 0.8077 18  3 M (0.83333333 0.16666667) *
#>         19) V20>=0.8077 8  0 R (0.00000000 1.00000000) *
#>      5) V36>=0.4926 35  9 R (0.25714286 0.74285714)  
#>       10) V39>=0.57105 9  1 M (0.88888889 0.11111111) *
#>       11) V39< 0.57105 26  1 R (0.03846154 0.96153846) *
#>    3) V48< 0.05195 40  5 R (0.12500000 0.87500000)  
#>      6) V29< 0.4323 9  4 M (0.55555556 0.44444444) *
#>      7) V29>=0.4323 31  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V52>=0.00935 81 25 M (0.69135802 0.30864198)  
#>      4) V12>=0.1271 64 12 M (0.81250000 0.18750000)  
#>        8) V60>=0.00285 50  3 M (0.94000000 0.06000000) *
#>        9) V60< 0.00285 14  5 R (0.35714286 0.64285714) *
#>      5) V12< 0.1271 17  4 R (0.23529412 0.76470588) *
#>    3) V52< 0.00935 58 13 R (0.22413793 0.77586207)  
#>      6) V51>=0.01445 9  2 M (0.77777778 0.22222222) *
#>      7) V51< 0.01445 49  6 R (0.12244898 0.87755102)  
#>       14) V9>=0.19205 7  2 M (0.71428571 0.28571429) *
#>       15) V9< 0.19205 42  1 R (0.02380952 0.97619048) *
#> 
#> 
#> $weights
#>  [1] 0.6665923 0.9451797 0.4986118 0.7358622 0.8752764 1.0418777 0.7320963
#>  [8] 0.8608312 0.8684545 0.8279268
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.7289930 5.3237159
#>   [2,] 2.5716577 5.4810512
#>   [3,] 2.9270893 5.1256196
#>   [4,] 2.2423427 5.8103662
#>   [5,] 1.4437915 6.6089174
#>   [6,] 0.8608312 7.1918777
#>   [7,] 0.8684545 7.1842544
#>   [8,] 1.7032032 6.3495057
#>   [9,] 2.3697955 5.6829134
#>  [10,] 0.0000000 8.0527089
#>  [11,] 2.8615853 5.1911236
#>  [12,] 3.1347962 4.9179127
#>  [13,] 0.0000000 8.0527089
#>  [14,] 1.6111386 6.4415703
#>  [15,] 1.9171541 6.1355547
#>  [16,] 2.0404805 6.0122284
#>  [17,] 1.3986885 6.6540204
#>  [18,] 1.8060110 6.2466979
#>  [19,] 0.8752764 7.1774325
#>  [20,] 1.4437915 6.6089174
#>  [21,] 2.2266153 5.8260936
#>  [22,] 3.1018917 4.9508172
#>  [23,] 0.8608312 7.1918777
#>  [24,] 0.8608312 7.1918777
#>  [25,] 0.0000000 8.0527089
#>  [26,] 0.0000000 8.0527089
#>  [27,] 1.7084700 6.3442389
#>  [28,] 0.0000000 8.0527089
#>  [29,] 3.1448062 4.9079027
#>  [30,] 2.6045621 5.4481467
#>  [31,] 1.5418687 6.5108402
#>  [32,] 1.9027090 6.1499999
#>  [33,] 1.8060110 6.2466979
#>  [34,] 0.8752764 7.1774325
#>  [35,] 1.9870575 6.0656514
#>  [36,] 0.8684545 7.1842544
#>  [37,] 0.4986118 7.5540971
#>  [38,] 1.8136342 6.2390746
#>  [39,] 2.3958780 5.6568309
#>  [40,] 2.4802265 5.5724824
#>  [41,] 0.0000000 8.0527089
#>  [42,] 0.8608312 7.1918777
#>  [43,] 1.5274235 6.5252854
#>  [44,] 1.5274235 6.5252854
#>  [45,] 1.5274235 6.5252854
#>  [46,] 1.6963813 6.3563276
#>  [47,] 0.8684545 7.1842544
#>  [48,] 0.8684545 7.1842544
#>  [49,] 1.0418777 7.0108312
#>  [50,] 0.0000000 8.0527089
#>  [51,] 1.0418777 7.0108312
#>  [52,] 1.3670663 6.6856426
#>  [53,] 0.8608312 7.1918777
#>  [54,] 0.8608312 7.1918777
#>  [55,] 1.6887580 6.3639509
#>  [56,] 2.2423427 5.8103662
#>  [57,] 1.7032032 6.3495057
#>  [58,] 1.3738882 6.6788207
#>  [59,] 2.5563184 5.4963905
#>  [60,] 2.0404805 6.0122284
#>  [61,] 2.2347194 5.8179895
#>  [62,] 2.3046227 5.7480861
#>  [63,] 1.3738882 6.6788207
#>  [64,] 2.4802265 5.5724824
#>  [65,] 2.4405663 5.6121426
#>  [66,] 0.8752764 7.1774325
#>  [67,] 1.5350468 6.5176621
#>  [68,] 8.0527089 0.0000000
#>  [69,] 5.5428727 2.5098362
#>  [70,] 5.7931891 2.2595198
#>  [71,] 4.7842159 3.2684930
#>  [72,] 6.6502544 1.4024545
#>  [73,] 4.9209687 3.1317402
#>  [74,] 7.3168467 0.7358622
#>  [75,] 6.7261703 1.3265385
#>  [76,] 4.8442434 3.2084654
#>  [77,] 8.0527089 0.0000000
#>  [78,] 7.0108312 1.0418777
#>  [79,] 6.4889199 1.5637890
#>  [80,] 6.8182349 1.2344740
#>  [81,] 6.1355547 1.9171541
#>  [82,] 7.0108312 1.0418777
#>  [83,] 5.8248114 2.2278975
#>  [84,] 8.0527089 0.0000000
#>  [85,] 6.1554086 1.8973003
#>  [86,] 6.1554086 1.8973003
#>  [87,] 7.3206126 0.7320963
#>  [88,] 6.5847504 1.4679585
#>  [89,] 5.6318546 2.4208543
#>  [90,] 8.0527089 0.0000000
#>  [91,] 7.1075291 0.9451797
#>  [92,] 6.6089174 1.4437915
#>  [93,] 7.2247821 0.8279268
#>  [94,] 7.3861166 0.6665923
#>  [95,] 6.8182349 1.2344740
#>  [96,] 6.3166012 1.7361076
#>  [97,] 8.0527089 0.0000000
#>  [98,] 8.0527089 0.0000000
#>  [99,] 5.7568237 2.2958852
#> [100,] 7.3206126 0.7320963
#> [101,] 7.5540971 0.4986118
#> [102,] 6.8182349 1.2344740
#> [103,] 6.2796024 1.7731065
#> [104,] 6.3716669 1.6810419
#> [105,] 6.6089174 1.4437915
#> [106,] 4.9521964 3.1005125
#> [107,] 6.4597814 1.5929275
#> [108,] 7.1842544 0.8684545
#> [109,] 7.0108312 1.0418777
#> [110,] 6.5122194 1.5404895
#> [111,] 5.9181582 2.1345507
#> [112,] 5.7094740 2.3432349
#> [113,] 5.5069784 2.5457305
#> [114,] 5.4111479 2.6415610
#> [115,] 5.7855659 2.2671430
#> [116,] 6.1423767 1.9103322
#> [117,] 5.8653391 2.1873698
#> [118,] 6.4521581 1.6005508
#> [119,] 7.3168467 0.7358622
#> [120,] 5.7763572 2.2763517
#> [121,] 6.6932659 1.3594430
#> [122,] 6.3639509 1.6887580
#> [123,] 7.1918777 0.8608312
#> [124,] 8.0527089 0.0000000
#> [125,] 8.0527089 0.0000000
#> [126,] 7.3168467 0.7358622
#> [127,] 6.4889199 1.5637890
#> [128,] 6.7261703 1.3265385
#> [129,] 6.1829044 1.8698045
#> [130,] 8.0527089 0.0000000
#> [131,] 6.1829044 1.8698045
#> [132,] 6.6089174 1.4437915
#> [133,] 6.8182349 1.2344740
#> [134,] 6.8182349 1.2344740
#> [135,] 6.8182349 1.2344740
#> [136,] 7.5540971 0.4986118
#> [137,] 8.0527089 0.0000000
#> [138,] 7.5540971 0.4986118
#> [139,] 5.0442609 3.0084480
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.33889130 0.66110870
#>   [2,] 0.31935312 0.68064688
#>   [3,] 0.36349126 0.63650874
#>   [4,] 0.27845818 0.72154182
#>   [5,] 0.17929265 0.82070735
#>   [6,] 0.10689958 0.89310042
#>   [7,] 0.10784625 0.89215375
#>   [8,] 0.21150686 0.78849314
#>   [9,] 0.29428550 0.70571450
#>  [10,] 0.00000000 1.00000000
#>  [11,] 0.35535685 0.64464315
#>  [12,] 0.38928468 0.61071532
#>  [13,] 0.00000000 1.00000000
#>  [14,] 0.20007412 0.79992588
#>  [15,] 0.23807568 0.76192432
#>  [16,] 0.25339057 0.74660943
#>  [17,] 0.17369168 0.82630832
#>  [18,] 0.22427372 0.77572628
#>  [19,] 0.10869341 0.89130659
#>  [20,] 0.17929265 0.82070735
#>  [21,] 0.27650513 0.72349487
#>  [22,] 0.38519854 0.61480146
#>  [23,] 0.10689958 0.89310042
#>  [24,] 0.10689958 0.89310042
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.21216090 0.78783910
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.39052774 0.60947226
#>  [30,] 0.32343925 0.67656075
#>  [31,] 0.19147205 0.80852795
#>  [32,] 0.23628185 0.76371815
#>  [33,] 0.22427372 0.77572628
#>  [34,] 0.10869341 0.89130659
#>  [35,] 0.24675640 0.75324360
#>  [36,] 0.10784625 0.89215375
#>  [37,] 0.06191851 0.93808149
#>  [38,] 0.22522039 0.77477961
#>  [39,] 0.29752448 0.70247552
#>  [40,] 0.30799903 0.69200097
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.10689958 0.89310042
#>  [43,] 0.18967822 0.81032178
#>  [44,] 0.18967822 0.81032178
#>  [45,] 0.18967822 0.81032178
#>  [46,] 0.21065970 0.78934030
#>  [47,] 0.10784625 0.89215375
#>  [48,] 0.10784625 0.89215375
#>  [49,] 0.12938227 0.87061773
#>  [50,] 0.00000000 1.00000000
#>  [51,] 0.12938227 0.87061773
#>  [52,] 0.16976477 0.83023523
#>  [53,] 0.10689958 0.89310042
#>  [54,] 0.10689958 0.89310042
#>  [55,] 0.20971303 0.79028697
#>  [56,] 0.27845818 0.72154182
#>  [57,] 0.21150686 0.78849314
#>  [58,] 0.17061193 0.82938807
#>  [59,] 0.31744825 0.68255175
#>  [60,] 0.25339057 0.74660943
#>  [61,] 0.27751151 0.72248849
#>  [62,] 0.28619223 0.71380777
#>  [63,] 0.17061193 0.82938807
#>  [64,] 0.30799903 0.69200097
#>  [65,] 0.30307395 0.69692605
#>  [66,] 0.10869341 0.89130659
#>  [67,] 0.19062489 0.80937511
#>  [68,] 1.00000000 0.00000000
#>  [69,] 0.68832399 0.31167601
#>  [70,] 0.71940874 0.28059126
#>  [71,] 0.59411261 0.40588739
#>  [72,] 0.82584066 0.17415934
#>  [73,] 0.61109482 0.38890518
#>  [74,] 0.90861930 0.09138070
#>  [75,] 0.83526804 0.16473196
#>  [76,] 0.60156694 0.39843306
#>  [77,] 1.00000000 0.00000000
#>  [78,] 0.87061773 0.12938227
#>  [79,] 0.80580585 0.19419415
#>  [80,] 0.84670078 0.15329922
#>  [81,] 0.76192432 0.23807568
#>  [82,] 0.87061773 0.12938227
#>  [83,] 0.72333565 0.27666435
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.76438981 0.23561019
#>  [86,] 0.76438981 0.23561019
#>  [87,] 0.90908696 0.09091304
#>  [88,] 0.81770625 0.18229375
#>  [89,] 0.69937393 0.30062607
#>  [90,] 1.00000000 0.00000000
#>  [91,] 0.88262586 0.11737414
#>  [92,] 0.82070735 0.17929265
#>  [93,] 0.89718655 0.10281345
#>  [94,] 0.91722136 0.08277864
#>  [95,] 0.84670078 0.15329922
#>  [96,] 0.78440700 0.21559300
#>  [97,] 1.00000000 0.00000000
#>  [98,] 1.00000000 0.00000000
#>  [99,] 0.71489281 0.28510719
#> [100,] 0.90908696 0.09091304
#> [101,] 0.93808149 0.06191851
#> [102,] 0.84670078 0.15329922
#> [103,] 0.77981242 0.22018758
#> [104,] 0.79124516 0.20875484
#> [105,] 0.82070735 0.17929265
#> [106,] 0.61497273 0.38502727
#> [107,] 0.80218737 0.19781263
#> [108,] 0.89215375 0.10784625
#> [109,] 0.87061773 0.12938227
#> [110,] 0.80869922 0.19130078
#> [111,] 0.73492762 0.26507238
#> [112,] 0.70901284 0.29098716
#> [113,] 0.68386657 0.31613343
#> [114,] 0.67196616 0.32803384
#> [115,] 0.71846207 0.28153793
#> [116,] 0.76277148 0.23722852
#> [117,] 0.72836845 0.27163155
#> [118,] 0.80124070 0.19875930
#> [119,] 0.90861930 0.09138070
#> [120,] 0.71731852 0.28268148
#> [121,] 0.83118190 0.16881810
#> [122,] 0.79028697 0.20971303
#> [123,] 0.89310042 0.10689958
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 0.90861930 0.09138070
#> [127,] 0.80580585 0.19419415
#> [128,] 0.83526804 0.16473196
#> [129,] 0.76780429 0.23219571
#> [130,] 1.00000000 0.00000000
#> [131,] 0.76780429 0.23219571
#> [132,] 0.82070735 0.17929265
#> [133,] 0.84670078 0.15329922
#> [134,] 0.84670078 0.15329922
#> [135,] 0.84670078 0.15329922
#> [136,] 0.93808149 0.06191851
#> [137,] 1.00000000 0.00000000
#> [138,] 0.93808149 0.06191851
#> [139,] 0.62640547 0.37359453
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 1.1177959 0.0000000 6.7203455 4.9453211 2.2363561 0.0000000 1.8366749 2.1676988 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 0.0000000 1.8735138 1.6222257 0.0000000 0.0000000 5.0729915 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 2.2472967 3.7822124 1.5752722 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 2.6753692 0.0000000 0.0000000 0.0000000 0.0000000 9.8260208 1.6080364 0.7882770 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 3.1792383 6.7602563 0.0000000 0.0000000 0.0000000 1.5712933 0.0000000 1.9835373 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 4.9165752 7.9147655 0.0000000 0.0000000 2.0257379 7.7032912 4.3438047 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 1.1099695 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 1.4940729 0.9912872 1.6212081 4.2895545 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V36       V48       V51        V4       V11       V23       V12       V47 
#> 9.8260208 7.9147655 7.7032912 6.7602563 6.7203455 5.0729915 4.9453211 4.9165752 
#>       V52        V9       V28       V39       V31       V27       V13       V16 
#> 4.3438047 4.2895545 3.7822124 3.1792383 2.6753692 2.2472967 2.2363561 2.1676988 
#>       V50       V45        V2       V15       V20        V8       V37       V29 
#> 2.0257379 1.9835373 1.8735138 1.8366749 1.6222257 1.6212081 1.6080364 1.5752722 
#>       V43       V60        V1       V54        V7       V38       V10       V14 
#> 1.5712933 1.4940729 1.1177959 1.1099695 0.9912872 0.7882770 0.0000000 0.0000000 
#>       V17       V18       V19       V21       V22       V24       V25       V26 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>        V3       V30       V32       V33       V34       V35       V40       V41 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V42       V44       V46       V49        V5       V53       V55       V56 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V57       V58       V59        V6 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913