Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 M (0.60431655 0.39568345)  
#>    2) V11>=0.1573 100 21 M (0.79000000 0.21000000)  
#>      4) V48>=0.0755 68  3 M (0.95588235 0.04411765) *
#>      5) V48< 0.0755 32 14 R (0.43750000 0.56250000)  
#>       10) V22>=0.68675 16  3 M (0.81250000 0.18750000) *
#>       11) V22< 0.68675 16  1 R (0.06250000 0.93750000) *
#>    3) V11< 0.1573 39  5 R (0.12820513 0.87179487)  
#>      6) V10>=0.13365 7  3 M (0.57142857 0.42857143) *
#>      7) V10< 0.13365 32  1 R (0.03125000 0.96875000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V12>=0.19455 75 20 M (0.73333333 0.26666667)  
#>      4) V27>=0.7993 38  0 M (1.00000000 0.00000000) *
#>      5) V27< 0.7993 37 17 R (0.45945946 0.54054054)  
#>       10) V32< 0.49505 20  4 M (0.80000000 0.20000000)  
#>         20) V39>=0.1859 13  0 M (1.00000000 0.00000000) *
#>         21) V39< 0.1859 7  3 R (0.42857143 0.57142857) *
#>       11) V32>=0.49505 17  1 R (0.05882353 0.94117647) *
#>    3) V12< 0.19455 64 11 R (0.17187500 0.82812500)  
#>      6) V20>=0.8321 16  7 M (0.56250000 0.43750000) *
#>      7) V20< 0.8321 48  2 R (0.04166667 0.95833333) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 69 M (0.50359712 0.49640288)  
#>   2) V11>=0.19615 84 22 M (0.73809524 0.26190476)  
#>     4) V47>=0.0639 72 11 M (0.84722222 0.15277778)  
#>       8) V37< 0.48 54  0 M (1.00000000 0.00000000) *
#>       9) V37>=0.48 18  7 R (0.38888889 0.61111111) *
#>     5) V47< 0.0639 12  1 R (0.08333333 0.91666667) *
#>   3) V11< 0.19615 55  8 R (0.14545455 0.85454545)  
#>     6) V1>=0.0321 8  2 M (0.75000000 0.25000000) *
#>     7) V1< 0.0321 47  2 R (0.04255319 0.95744681) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V47>=0.10475 66 10 M (0.84848485 0.15151515)  
#>      4) V41< 0.66555 57  3 M (0.94736842 0.05263158) *
#>      5) V41>=0.66555 9  2 R (0.22222222 0.77777778) *
#>    3) V47< 0.10475 73 17 R (0.23287671 0.76712329)  
#>      6) V19>=0.92575 7  0 M (1.00000000 0.00000000) *
#>      7) V19< 0.92575 66 10 R (0.15151515 0.84848485)  
#>       14) V51>=0.0163 13  5 M (0.61538462 0.38461538) *
#>       15) V51< 0.0163 53  2 R (0.03773585 0.96226415) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V12>=0.2244 86 18 M (0.79069767 0.20930233)  
#>     4) V47>=0.06675 77 11 M (0.85714286 0.14285714) *
#>     5) V47< 0.06675 9  2 R (0.22222222 0.77777778) *
#>   3) V12< 0.2244 53 13 R (0.24528302 0.75471698)  
#>     6) V19>=0.7815 14  3 M (0.78571429 0.21428571) *
#>     7) V19< 0.7815 39  2 R (0.05128205 0.94871795) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 67 M (0.51798561 0.48201439)  
#>   2) V36< 0.43825 75 17 M (0.77333333 0.22666667)  
#>     4) V42>=0.10015 65  7 M (0.89230769 0.10769231) *
#>     5) V42< 0.10015 10  0 R (0.00000000 1.00000000) *
#>   3) V36>=0.43825 64 14 R (0.21875000 0.78125000)  
#>     6) V4>=0.0606 19  7 M (0.63157895 0.36842105) *
#>     7) V4< 0.0606 45  2 R (0.04444444 0.95555556) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V48>=0.08475 58 16 M (0.72413793 0.27586207)  
#>      4) V34>=0.07815 48  6 M (0.87500000 0.12500000)  
#>        8) V41< 0.6322 40  1 M (0.97500000 0.02500000) *
#>        9) V41>=0.6322 8  3 R (0.37500000 0.62500000) *
#>      5) V34< 0.07815 10  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.08475 81 19 R (0.23456790 0.76543210)  
#>      6) V31< 0.3838 29 12 M (0.58620690 0.41379310)  
#>       12) V38>=0.30135 19  2 M (0.89473684 0.10526316) *
#>       13) V38< 0.30135 10  0 R (0.00000000 1.00000000) *
#>      7) V31>=0.3838 52  2 R (0.03846154 0.96153846) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V11>=0.198 76 20 M (0.73684211 0.26315789)  
#>      4) V16< 0.67815 56  6 M (0.89285714 0.10714286)  
#>        8) V47>=0.06785 49  2 M (0.95918367 0.04081633) *
#>        9) V47< 0.06785 7  3 R (0.42857143 0.57142857) *
#>      5) V16>=0.67815 20  6 R (0.30000000 0.70000000)  
#>       10) V27< 0.49675 7  1 M (0.85714286 0.14285714) *
#>       11) V27>=0.49675 13  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.198 63 12 R (0.19047619 0.80952381)  
#>      6) V46>=0.16875 11  3 M (0.72727273 0.27272727) *
#>      7) V46< 0.16875 52  4 R (0.07692308 0.92307692) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.4604317 0.5395683)  
#>    2) V37< 0.2468 49 10 M (0.7959184 0.2040816)  
#>      4) V6>=0.0924 37  0 M (1.0000000 0.0000000) *
#>      5) V6< 0.0924 12  2 R (0.1666667 0.8333333) *
#>    3) V37>=0.2468 90 25 R (0.2777778 0.7222222)  
#>      6) V46>=0.1709 23  8 M (0.6521739 0.3478261)  
#>       12) V10>=0.12205 14  0 M (1.0000000 0.0000000) *
#>       13) V10< 0.12205 9  1 R (0.1111111 0.8888889) *
#>      7) V46< 0.1709 67 10 R (0.1492537 0.8507463)  
#>       14) V36< 0.52295 29 10 R (0.3448276 0.6551724)  
#>         28) V42>=0.32315 7  0 M (1.0000000 0.0000000) *
#>         29) V42< 0.32315 22  3 R (0.1363636 0.8636364) *
#>       15) V36>=0.52295 38  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 55 M (0.60431655 0.39568345)  
#>   2) V36< 0.52905 98 22 M (0.77551020 0.22448980)  
#>     4) V21>=0.4968 81  9 M (0.88888889 0.11111111)  
#>       8) V39>=0.10305 74  4 M (0.94594595 0.05405405) *
#>       9) V39< 0.10305 7  2 R (0.28571429 0.71428571) *
#>     5) V21< 0.4968 17  4 R (0.23529412 0.76470588) *
#>   3) V36>=0.52905 41  8 R (0.19512195 0.80487805)  
#>     6) V27>=0.8444 9  1 M (0.88888889 0.11111111) *
#>     7) V27< 0.8444 32  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.9527094 0.9403133 0.9104415 0.6515645 0.6853443 0.7191001 0.9120858
#>  [8] 0.8030403 0.8700509 0.7155391
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.5163597 5.6438297
#>   [2,] 1.5185794 6.6416100
#>   [3,] 2.7715481 5.3886413
#>   [4,] 1.6594135 6.5007759
#>   [5,] 0.0000000 8.1601894
#>   [6,] 2.0862037 6.0739857
#>   [7,] 1.5918779 6.5683115
#>   [8,] 1.5918779 6.5683115
#>   [9,] 2.3467251 5.8134644
#>  [10,] 2.8590493 5.3011401
#>  [11,] 2.6134046 5.5467848
#>  [12,] 2.2827505 5.8774389
#>  [13,] 1.6718095 6.4883799
#>  [14,] 2.4410941 5.7190954
#>  [15,] 2.6258006 5.5343888
#>  [16,] 0.9120858 7.2481036
#>  [17,] 1.5636503 6.5965391
#>  [18,] 0.8030403 7.3571491
#>  [19,] 2.9900345 5.1701549
#>  [20,] 2.1399492 6.0202402
#>  [21,] 1.6718095 6.4883799
#>  [22,] 2.3411968 5.8189926
#>  [23,] 2.3074170 5.8527725
#>  [24,] 2.2473504 5.9128390
#>  [25,] 3.0726930 5.0874964
#>  [26,] 0.0000000 8.1601894
#>  [27,] 0.6515645 7.5086249
#>  [28,] 0.0000000 8.1601894
#>  [29,] 0.8030403 7.3571491
#>  [30,] 0.9527094 7.2074800
#>  [31,] 2.3571539 5.8030355
#>  [32,] 2.3988262 5.7613632
#>  [33,] 0.7155391 7.4446503
#>  [34,] 0.0000000 8.1601894
#>  [35,] 0.7191001 7.4410893
#>  [36,] 1.6682485 6.4919409
#>  [37,] 0.7155391 7.4446503
#>  [38,] 1.6682485 6.4919409
#>  [39,] 0.0000000 8.1601894
#>  [40,] 0.9120858 7.2481036
#>  [41,] 1.5957859 6.5644035
#>  [42,] 1.4546048 6.7055846
#>  [43,] 0.7191001 7.4410893
#>  [44,] 0.7191001 7.4410893
#>  [45,] 1.5221405 6.6380489
#>  [46,] 0.6853443 7.4748451
#>  [47,] 2.5783671 5.5818223
#>  [48,] 1.5957859 6.5644035
#>  [49,] 1.6558524 6.5043370
#>  [50,] 0.7155391 7.4446503
#>  [51,] 2.3233740 5.8368154
#>  [52,] 1.6042739 6.5559155
#>  [53,] 0.0000000 8.1601894
#>  [54,] 1.5855900 6.5745994
#>  [55,] 1.6594135 6.5007759
#>  [56,] 0.7191001 7.4410893
#>  [57,] 1.6311860 6.5290034
#>  [58,] 6.5325645 1.6276249
#>  [59,] 7.4446503 0.7155391
#>  [60,] 5.9230349 2.2371545
#>  [61,] 6.3077902 1.8523992
#>  [62,] 6.5710383 1.5891511
#>  [63,] 5.6943526 2.4658368
#>  [64,] 6.2671667 1.8930227
#>  [65,] 6.4044397 1.7557497
#>  [66,] 5.6858646 2.4743248
#>  [67,] 7.3571491 0.8030403
#>  [68,] 6.2970385 1.8631509
#>  [69,] 7.3571491 0.8030403
#>  [70,] 6.6718047 1.4883847
#>  [71,] 7.2901385 0.8700509
#>  [72,] 8.1601894 0.0000000
#>  [73,] 7.2481036 0.9120858
#>  [74,] 6.5627593 1.5974302
#>  [75,] 4.1977129 3.9624765
#>  [76,] 6.3780527 1.7821367
#>  [77,] 7.2901385 0.8700509
#>  [78,] 7.2901385 0.8700509
#>  [79,] 7.2481036 0.9120858
#>  [80,] 6.4168357 1.7433537
#>  [81,] 7.3571491 0.8030403
#>  [82,] 7.2481036 0.9120858
#>  [83,] 7.4748451 0.6853443
#>  [84,] 5.8488645 2.3113249
#>  [85,] 6.5342088 1.6259806
#>  [86,] 5.5814994 2.5786900
#>  [87,] 5.8488645 2.3113249
#>  [88,] 6.5745994 1.5855900
#>  [89,] 4.8913123 3.2688771
#>  [90,] 7.2198761 0.9403133
#>  [91,] 5.6562257 2.5039637
#>  [92,] 6.3796970 1.7804924
#>  [93,] 6.5043370 1.6558524
#>  [94,] 8.1601894 0.0000000
#>  [95,] 6.5965391 1.5636503
#>  [96,] 5.1501893 3.0100001
#>  [97,] 6.3780527 1.7821367
#>  [98,] 5.7311685 2.4290209
#>  [99,] 5.6221230 2.5380664
#> [100,] 5.5346217 2.6255677
#> [101,] 6.5306477 1.6295417
#> [102,] 7.2497479 0.9104415
#> [103,] 5.8151087 2.3450808
#> [104,] 7.2497479 0.9104415
#> [105,] 6.5306477 1.6295417
#> [106,] 7.2901385 0.8700509
#> [107,] 5.6562257 2.5039637
#> [108,] 5.6307250 2.5294644
#> [109,] 5.8189926 2.3411968
#> [110,] 5.8790832 2.2811062
#> [111,] 6.5981834 1.5620060
#> [112,] 6.5306477 1.6295417
#> [113,] 5.7281324 2.4320570
#> [114,] 5.7264882 2.4337013
#> [115,] 6.5683115 1.5918779
#> [116,] 8.1601894 0.0000000
#> [117,] 8.1601894 0.0000000
#> [118,] 8.1601894 0.0000000
#> [119,] 8.1601894 0.0000000
#> [120,] 8.1601894 0.0000000
#> [121,] 8.1601894 0.0000000
#> [122,] 6.5745994 1.5855900
#> [123,] 6.5745994 1.5855900
#> [124,] 7.4446503 0.7155391
#> [125,] 6.5745994 1.5855900
#> [126,] 7.4446503 0.7155391
#> [127,] 7.4446503 0.7155391
#> [128,] 6.5745994 1.5855900
#> [129,] 5.7264882 2.4337013
#> [130,] 7.2901385 0.8700509
#> [131,] 7.2901385 0.8700509
#> [132,] 7.2901385 0.8700509
#> [133,] 8.1601894 0.0000000
#> [134,] 7.4446503 0.7155391
#> [135,] 7.4446503 0.7155391
#> [136,] 6.0739857 2.0862037
#> [137,] 6.7255502 1.4346392
#> [138,] 8.1601894 0.0000000
#> [139,] 5.6625136 2.4976758
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.30837026 0.69162974
#>   [2,] 0.18609610 0.81390390
#>   [3,] 0.33964262 0.66035738
#>   [4,] 0.20335478 0.79664522
#>   [5,] 0.00000000 1.00000000
#>   [6,] 0.25565629 0.74434371
#>   [7,] 0.19507854 0.80492146
#>   [8,] 0.19507854 0.80492146
#>   [9,] 0.28758218 0.71241782
#>  [10,] 0.35036556 0.64963444
#>  [11,] 0.32026274 0.67973726
#>  [12,] 0.27974234 0.72025766
#>  [13,] 0.20487386 0.79512614
#>  [14,] 0.29914674 0.70085326
#>  [15,] 0.32178182 0.67821818
#>  [16,] 0.11177263 0.88822737
#>  [17,] 0.19161937 0.80838063
#>  [18,] 0.09840952 0.90159048
#>  [19,] 0.36641729 0.63358271
#>  [20,] 0.26224259 0.73775741
#>  [21,] 0.20487386 0.79512614
#>  [22,] 0.28690471 0.71309529
#>  [23,] 0.28276512 0.71723488
#>  [24,] 0.27540419 0.72459581
#>  [25,] 0.37654677 0.62345323
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.07984674 0.92015326
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.09840952 0.90159048
#>  [30,] 0.11675089 0.88324911
#>  [31,] 0.28886019 0.71113981
#>  [32,] 0.29396697 0.70603303
#>  [33,] 0.08768658 0.91231342
#>  [34,] 0.00000000 1.00000000
#>  [35,] 0.08812297 0.91187703
#>  [36,] 0.20443747 0.79556253
#>  [37,] 0.08768658 0.91231342
#>  [38,] 0.20443747 0.79556253
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.11177263 0.88822737
#>  [41,] 0.19555745 0.80444255
#>  [42,] 0.17825626 0.82174374
#>  [43,] 0.08812297 0.91187703
#>  [44,] 0.08812297 0.91187703
#>  [45,] 0.18653249 0.81346751
#>  [46,] 0.08398633 0.91601367
#>  [47,] 0.31596902 0.68403098
#>  [48,] 0.19555745 0.80444255
#>  [49,] 0.20291838 0.79708162
#>  [50,] 0.08768658 0.91231342
#>  [51,] 0.28472060 0.71527940
#>  [52,] 0.19659763 0.80340237
#>  [53,] 0.00000000 1.00000000
#>  [54,] 0.19430799 0.80569201
#>  [55,] 0.20335478 0.79664522
#>  [56,] 0.08812297 0.91187703
#>  [57,] 0.19989560 0.80010440
#>  [58,] 0.80054079 0.19945921
#>  [59,] 0.91231342 0.08768658
#>  [60,] 0.72584527 0.27415473
#>  [61,] 0.77299557 0.22700443
#>  [62,] 0.80525562 0.19474438
#>  [63,] 0.69782113 0.30217887
#>  [64,] 0.76801730 0.23198270
#>  [65,] 0.78483959 0.21516041
#>  [66,] 0.69678096 0.30321904
#>  [67,] 0.90159048 0.09840952
#>  [68,] 0.77167798 0.22832202
#>  [69,] 0.90159048 0.09840952
#>  [70,] 0.81760415 0.18239585
#>  [71,] 0.89337859 0.10662141
#>  [72,] 1.00000000 0.00000000
#>  [73,] 0.88822737 0.11177263
#>  [74,] 0.80424104 0.19575896
#>  [75,] 0.51441366 0.48558634
#>  [76,] 0.78160596 0.21839404
#>  [77,] 0.89337859 0.10662141
#>  [78,] 0.89337859 0.10662141
#>  [79,] 0.88822737 0.11177263
#>  [80,] 0.78635868 0.21364132
#>  [81,] 0.90159048 0.09840952
#>  [82,] 0.88822737 0.11177263
#>  [83,] 0.91601367 0.08398633
#>  [84,] 0.71675597 0.28324403
#>  [85,] 0.80074229 0.19925771
#>  [86,] 0.68399140 0.31600860
#>  [87,] 0.71675597 0.28324403
#>  [88,] 0.80569201 0.19430799
#>  [89,] 0.59941161 0.40058839
#>  [90,] 0.88476820 0.11523180
#>  [91,] 0.69314883 0.30685117
#>  [92,] 0.78180746 0.21819254
#>  [93,] 0.79708162 0.20291838
#>  [94,] 1.00000000 0.00000000
#>  [95,] 0.80838063 0.19161937
#>  [96,] 0.63113600 0.36886400
#>  [97,] 0.78160596 0.21839404
#>  [98,] 0.70233277 0.29766723
#>  [99,] 0.68896967 0.31103033
#> [100,] 0.67824672 0.32175328
#> [101,] 0.80030590 0.19969410
#> [102,] 0.88842887 0.11157113
#> [103,] 0.71261932 0.28738068
#> [104,] 0.88842887 0.11157113
#> [105,] 0.80030590 0.19969410
#> [106,] 0.89337859 0.10662141
#> [107,] 0.69314883 0.30685117
#> [108,] 0.69002381 0.30997619
#> [109,] 0.71309529 0.28690471
#> [110,] 0.72045916 0.27954084
#> [111,] 0.80858213 0.19141787
#> [112,] 0.80030590 0.19969410
#> [113,] 0.70196072 0.29803928
#> [114,] 0.70175922 0.29824078
#> [115,] 0.80492146 0.19507854
#> [116,] 1.00000000 0.00000000
#> [117,] 1.00000000 0.00000000
#> [118,] 1.00000000 0.00000000
#> [119,] 1.00000000 0.00000000
#> [120,] 1.00000000 0.00000000
#> [121,] 1.00000000 0.00000000
#> [122,] 0.80569201 0.19430799
#> [123,] 0.80569201 0.19430799
#> [124,] 0.91231342 0.08768658
#> [125,] 0.80569201 0.19430799
#> [126,] 0.91231342 0.08768658
#> [127,] 0.91231342 0.08768658
#> [128,] 0.80569201 0.19430799
#> [129,] 0.70175922 0.29824078
#> [130,] 0.89337859 0.10662141
#> [131,] 0.89337859 0.10662141
#> [132,] 0.89337859 0.10662141
#> [133,] 1.00000000 0.00000000
#> [134,] 0.91231342 0.08768658
#> [135,] 0.91231342 0.08768658
#> [136,] 0.74434371 0.25565629
#> [137,] 0.82419045 0.17580955
#> [138,] 1.00000000 0.00000000
#> [139,] 0.69391938 0.30608062
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#>  1.525164  2.625640 14.978675  8.284915  0.000000  0.000000  0.000000  2.036543 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#>  0.000000  0.000000  3.318083  0.000000  1.498660  2.102858  2.099059  0.000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#>  0.000000  0.000000  0.000000  5.780517  0.000000  0.000000  0.000000  0.000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#>  2.494371  2.324046  0.000000  2.829545  0.000000  7.983792  5.875788  2.342252 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#>  1.660618  1.621445  0.000000  2.375631  4.116786  0.000000  0.000000  0.000000 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#>  3.354688  8.634048  6.345097  0.000000  0.000000  0.000000  1.111251  0.000000 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  2.680530 
#>       V60        V7        V8        V9 
#>  0.000000  0.000000  0.000000  0.000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 82 57 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V11       V47       V12       V36       V48       V37       V27       V42 
#> 14.978675  8.634048  8.284915  7.983792  6.345097  5.875788  5.780517  4.116786 
#>       V46       V19       V34        V6       V10       V31       V41       V38 
#>  3.354688  3.318083  2.829545  2.680530  2.625640  2.494371  2.375631  2.342252 
#>       V32       V21       V22       V16       V39        V4        V1       V20 
#>  2.324046  2.102858  2.099059  2.036543  1.660618  1.621445  1.525164  1.498660 
#>       V51       V13       V14       V15       V17       V18        V2       V23 
#>  1.111251  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V24       V25       V26       V28       V29        V3       V30       V33 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V35       V40       V43       V44       V45       V49        V5       V50 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#>  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000  0.000000 
#>       V60        V7        V8        V9 
#>  0.000000  0.000000  0.000000  0.000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623