Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V49>=0.0368 76 16 M (0.78947368 0.21052632)  
#>      4) V37< 0.5062 55  3 M (0.94545455 0.05454545) *
#>      5) V37>=0.5062 21  8 R (0.38095238 0.61904762)  
#>       10) V54< 0.01085 8  0 M (1.00000000 0.00000000) *
#>       11) V54>=0.01085 13  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.0368 63 14 R (0.22222222 0.77777778)  
#>      6) V22>=0.8118 15  5 M (0.66666667 0.33333333) *
#>      7) V22< 0.8118 48  4 R (0.08333333 0.91666667)  
#>       14) V48>=0.07515 7  3 M (0.57142857 0.42857143) *
#>       15) V48< 0.07515 41  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 R (0.4316547 0.5683453)  
#>    2) V46>=0.07485 89 35 M (0.6067416 0.3932584)  
#>      4) V27>=0.8167 39  4 M (0.8974359 0.1025641)  
#>        8) V11>=0.1742 32  0 M (1.0000000 0.0000000) *
#>        9) V11< 0.1742 7  3 R (0.4285714 0.5714286) *
#>      5) V27< 0.8167 50 19 R (0.3800000 0.6200000)  
#>       10) V1>=0.03945 13  0 M (1.0000000 0.0000000) *
#>       11) V1< 0.03945 37  6 R (0.1621622 0.8378378)  
#>         22) V31< 0.4404 9  3 M (0.6666667 0.3333333) *
#>         23) V31>=0.4404 28  0 R (0.0000000 1.0000000) *
#>    3) V46< 0.07485 50  6 R (0.1200000 0.8800000)  
#>      6) V49>=0.0426 10  4 M (0.6000000 0.4000000) *
#>      7) V49< 0.0426 40  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V12>=0.22505 69 14 M (0.79710145 0.20289855)  
#>      4) V47>=0.06785 55  4 M (0.92727273 0.07272727) *
#>      5) V47< 0.06785 14  4 R (0.28571429 0.71428571) *
#>    3) V12< 0.22505 70 16 R (0.22857143 0.77142857)  
#>      6) V20>=0.5033 25 12 M (0.52000000 0.48000000)  
#>       12) V42>=0.08525 16  3 M (0.81250000 0.18750000) *
#>       13) V42< 0.08525 9  0 R (0.00000000 1.00000000) *
#>      7) V20< 0.5033 45  3 R (0.06666667 0.93333333) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V9>=0.1614 54 12 M (0.77777778 0.22222222)  
#>      4) V51>=0.01255 42  4 M (0.90476190 0.09523810)  
#>        8) V37< 0.6332 35  0 M (1.00000000 0.00000000) *
#>        9) V37>=0.6332 7  3 R (0.42857143 0.57142857) *
#>      5) V51< 0.01255 12  4 R (0.33333333 0.66666667) *
#>    3) V9< 0.1614 85 22 R (0.25882353 0.74117647)  
#>      6) V44>=0.24985 34 15 M (0.55882353 0.44117647)  
#>       12) V6< 0.1053 16  0 M (1.00000000 0.00000000) *
#>       13) V6>=0.1053 18  3 R (0.16666667 0.83333333) *
#>      7) V44< 0.24985 51  3 R (0.05882353 0.94117647) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 R (0.48920863 0.51079137)  
#>   2) V12>=0.22065 75 15 M (0.80000000 0.20000000)  
#>     4) V20< 0.92985 68  8 M (0.88235294 0.11764706) *
#>     5) V20>=0.92985 7  0 R (0.00000000 1.00000000) *
#>   3) V12< 0.22065 64  8 R (0.12500000 0.87500000)  
#>     6) V47>=0.12525 10  3 M (0.70000000 0.30000000) *
#>     7) V47< 0.12525 54  1 R (0.01851852 0.98148148) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 51 R (0.36690647 0.63309353)  
#>    2) V45>=0.2383 26  3 M (0.88461538 0.11538462) *
#>    3) V45< 0.2383 113 28 R (0.24778761 0.75221239)  
#>      6) V24>=0.6352 56 25 R (0.44642857 0.55357143)  
#>       12) V12>=0.15395 31  8 M (0.74193548 0.25806452)  
#>         24) V49>=0.032 23  2 M (0.91304348 0.08695652) *
#>         25) V49< 0.032 8  2 R (0.25000000 0.75000000) *
#>       13) V12< 0.15395 25  2 R (0.08000000 0.92000000) *
#>      7) V24< 0.6352 57  3 R (0.05263158 0.94736842) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 65 R (0.46762590 0.53237410)  
#>   2) V9>=0.1095 87 25 M (0.71264368 0.28735632)  
#>     4) V32>=0.09675 78 16 M (0.79487179 0.20512821)  
#>       8) V37< 0.4299 60  6 M (0.90000000 0.10000000) *
#>       9) V37>=0.4299 18  8 R (0.44444444 0.55555556) *
#>     5) V32< 0.09675 9  0 R (0.00000000 1.00000000) *
#>   3) V9< 0.1095 52  3 R (0.05769231 0.94230769) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V23>=0.79225 48 12 M (0.75000000 0.25000000)  
#>      4) V12>=0.07215 37  2 M (0.94594595 0.05405405) *
#>      5) V12< 0.07215 11  1 R (0.09090909 0.90909091) *
#>    3) V23< 0.79225 91 22 R (0.24175824 0.75824176)  
#>      6) V1>=0.0385 10  0 M (1.00000000 0.00000000) *
#>      7) V1< 0.0385 81 12 R (0.14814815 0.85185185)  
#>       14) V28>=0.9506 9  0 M (1.00000000 0.00000000) *
#>       15) V28< 0.9506 72  3 R (0.04166667 0.95833333) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V52>=0.00935 78 20 M (0.74358974 0.25641026)  
#>      4) V1>=0.0107 70 12 M (0.82857143 0.17142857)  
#>        8) V9>=0.15935 46  1 M (0.97826087 0.02173913) *
#>        9) V9< 0.15935 24 11 M (0.54166667 0.45833333)  
#>         18) V43>=0.24355 16  3 M (0.81250000 0.18750000) *
#>         19) V43< 0.24355 8  0 R (0.00000000 1.00000000) *
#>      5) V1< 0.0107 8  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.00935 61 18 R (0.29508197 0.70491803)  
#>      6) V45>=0.23475 12  2 M (0.83333333 0.16666667) *
#>      7) V45< 0.23475 49  8 R (0.16326531 0.83673469) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 R (0.39568345 0.60431655)  
#>    2) V37< 0.4751 84 36 M (0.57142857 0.42857143)  
#>      4) V49>=0.03625 41  8 M (0.80487805 0.19512195)  
#>        8) V16< 0.76295 34  3 M (0.91176471 0.08823529) *
#>        9) V16>=0.76295 7  2 R (0.28571429 0.71428571) *
#>      5) V49< 0.03625 43 15 R (0.34883721 0.65116279)  
#>       10) V34< 0.2077 11  1 M (0.90909091 0.09090909) *
#>       11) V34>=0.2077 32  5 R (0.15625000 0.84375000)  
#>         22) V42>=0.25 8  3 M (0.62500000 0.37500000) *
#>         23) V42< 0.25 24  0 R (0.00000000 1.00000000) *
#>    3) V37>=0.4751 55  7 R (0.12727273 0.87272727)  
#>      6) V27>=0.8777 8  2 M (0.75000000 0.25000000) *
#>      7) V27< 0.8777 47  1 R (0.02127660 0.97872340) *
#> 
#> 
#> $weights
#>  [1] 0.7346456 0.8441250 0.7352914 0.7612531 0.6537348 0.7923790 0.6217399
#>  [8] 0.8067998 0.8201429 0.8278488
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.4566469 5.1413135
#>   [2,] 2.2091692 5.3887912
#>   [3,] 2.2073669 5.3905935
#>   [4,] 3.0496365 4.5483239
#>   [5,] 2.1033235 5.4946369
#>   [6,] 2.4255209 5.1724395
#>   [7,] 2.3862162 5.2117442
#>   [8,] 2.9515207 4.6464397
#>   [9,] 1.6346486 5.9633118
#>  [10,] 0.0000000 7.5979604
#>  [11,] 1.5414455 6.0565149
#>  [12,] 1.4978598 6.1001006
#>  [13,] 2.4072652 5.1906952
#>  [14,] 1.4738777 6.1240827
#>  [15,] 0.8278488 6.7701116
#>  [16,] 2.3237475 5.2742129
#>  [17,] 1.4461139 6.1518465
#>  [18,] 0.7346456 6.8633148
#>  [19,] 1.3883805 6.2095799
#>  [20,] 3.2016238 4.3963366
#>  [21,] 1.5420913 6.0558691
#>  [22,] 2.1814053 5.4165551
#>  [23,] 2.1033235 5.4946369
#>  [24,] 1.6719737 5.9259867
#>  [25,] 1.5794164 6.0185440
#>  [26,] 1.4285397 6.1694206
#>  [27,] 0.7346456 6.8633148
#>  [28,] 0.0000000 7.5979604
#>  [29,] 0.8278488 6.7701116
#>  [30,] 1.3829930 6.2149674
#>  [31,] 1.6125219 5.9854384
#>  [32,] 3.0009024 4.5970580
#>  [33,] 3.0970386 4.5009218
#>  [34,] 1.5794164 6.0185440
#>  [35,] 2.1236719 5.4742885
#>  [36,] 0.8278488 6.7701116
#>  [37,] 1.5624944 6.0354660
#>  [38,] 0.8278488 6.7701116
#>  [39,] 1.6346486 5.9633118
#>  [40,] 0.0000000 7.5979604
#>  [41,] 1.6346486 5.9633118
#>  [42,] 0.7352914 6.8626690
#>  [43,] 0.7346456 6.8633148
#>  [44,] 0.8441250 6.7538354
#>  [45,] 0.8441250 6.7538354
#>  [46,] 1.3890263 6.2089341
#>  [47,] 2.3737750 5.2241854
#>  [48,] 0.6217399 6.9762205
#>  [49,] 0.8278488 6.7701116
#>  [50,] 1.4495887 6.1483717
#>  [51,] 1.6719737 5.9259867
#>  [52,] 0.7346456 6.8633148
#>  [53,] 1.3563856 6.2415748
#>  [54,] 1.3563856 6.2415748
#>  [55,] 1.3563856 6.2415748
#>  [56,] 1.5414455 6.0565149
#>  [57,] 1.4699371 6.1280233
#>  [58,] 2.7454118 4.8525486
#>  [59,] 2.7454118 4.8525486
#>  [60,] 2.3180027 5.2799577
#>  [61,] 2.3046596 5.2933008
#>  [62,] 1.5787706 6.0191898
#>  [63,] 2.1951803 5.4027801
#>  [64,] 2.0956177 5.5023427
#>  [65,] 1.4495887 6.1483717
#>  [66,] 2.9101233 4.6878371
#>  [67,] 2.8957025 4.7022579
#>  [68,] 0.7346456 6.8633148
#>  [69,] 7.5979604 0.0000000
#>  [70,] 5.1413135 2.4566469
#>  [71,] 4.4193651 3.1785953
#>  [72,] 5.3492777 2.2486827
#>  [73,] 6.2149674 1.3829930
#>  [74,] 5.5023427 2.0956177
#>  [75,] 5.4225884 2.1753720
#>  [76,] 6.2149674 1.3829930
#>  [77,] 6.2149674 1.3829930
#>  [78,] 6.1829725 1.4149879
#>  [79,] 5.3708425 2.2271179
#>  [80,] 6.7778175 0.8201429
#>  [81,] 6.0185440 1.5794164
#>  [82,] 6.0165644 1.5813960
#>  [83,] 6.0165644 1.5813960
#>  [84,] 6.7778175 0.8201429
#>  [85,] 7.5979604 0.0000000
#>  [86,] 6.0702899 1.5276705
#>  [87,] 6.1163768 1.4815836
#>  [88,] 6.7701116 0.8278488
#>  [89,] 6.0348202 1.5631402
#>  [90,] 5.3871186 2.2108417
#>  [91,] 4.6524730 2.9454874
#>  [92,] 5.3871186 2.2108417
#>  [93,] 7.5979604 0.0000000
#>  [94,] 6.0558691 1.5420913
#>  [95,] 7.5979604 0.0000000
#>  [96,] 7.5979604 0.0000000
#>  [97,] 5.2117442 2.3862162
#>  [98,] 6.2149674 1.3829930
#>  [99,] 6.8367073 0.7612531
#> [100,] 6.8055814 0.7923790
#> [101,] 6.8367073 0.7612531
#> [102,] 5.3077216 2.2902388
#> [103,] 5.5301066 2.0678538
#> [104,] 5.9614564 1.6365040
#> [105,] 4.5837937 3.0141667
#> [106,] 4.5704506 3.0275098
#> [107,] 5.3492777 2.2486827
#> [108,] 5.4081676 2.1897928
#> [109,] 6.2149674 1.3829930
#> [110,] 4.5803188 3.0176416
#> [111,] 6.0558691 1.5420913
#> [112,] 6.0165644 1.5813960
#> [113,] 5.9925824 1.6053780
#> [114,] 6.8626690 0.7352914
#> [115,] 6.1320955 1.4658649
#> [116,] 6.1560776 1.4418828
#> [117,] 6.2415748 1.3563856
#> [118,] 6.2415748 1.3563856
#> [119,] 6.8367073 0.7612531
#> [120,] 6.0088586 1.5891018
#> [121,] 6.2409290 1.3570314
#> [122,] 5.1724395 2.4255209
#> [123,] 6.8367073 0.7612531
#> [124,] 7.5979604 0.0000000
#> [125,] 5.9710176 1.6269428
#> [126,] 6.7778175 0.8201429
#> [127,] 7.5979604 0.0000000
#> [128,] 7.5979604 0.0000000
#> [129,] 6.7778175 0.8201429
#> [130,] 6.0165644 1.5813960
#> [131,] 6.0165644 1.5813960
#> [132,] 7.5979604 0.0000000
#> [133,] 6.8367073 0.7612531
#> [134,] 6.0354660 1.5624944
#> [135,] 7.5979604 0.0000000
#> [136,] 7.5979604 0.0000000
#> [137,] 7.5979604 0.0000000
#> [138,] 5.3948245 2.2031359
#> [139,] 6.0709357 1.5270247
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32332979 0.67667021
#>   [2,] 0.29075819 0.70924181
#>   [3,] 0.29052098 0.70947902
#>   [4,] 0.40137568 0.59862432
#>   [5,] 0.27682739 0.72317261
#>   [6,] 0.31923317 0.68076683
#>   [7,] 0.31406010 0.68593990
#>   [8,] 0.38846223 0.61153777
#>   [9,] 0.21514308 0.78485692
#>  [10,] 0.00000000 1.00000000
#>  [11,] 0.20287622 0.79712378
#>  [12,] 0.19713972 0.80286028
#>  [13,] 0.31683045 0.68316955
#>  [14,] 0.19398334 0.80601666
#>  [15,] 0.10895671 0.89104329
#>  [16,] 0.30583832 0.69416168
#>  [17,] 0.19032922 0.80967078
#>  [18,] 0.09668985 0.90331015
#>  [19,] 0.18273068 0.81726932
#>  [20,] 0.42137937 0.57862063
#>  [21,] 0.20296121 0.79703879
#>  [22,] 0.28710406 0.71289594
#>  [23,] 0.27682739 0.72317261
#>  [24,] 0.22005560 0.77994440
#>  [25,] 0.20787373 0.79212627
#>  [26,] 0.18801621 0.81198379
#>  [27,] 0.09668985 0.90331015
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.10895671 0.89104329
#>  [30,] 0.18202161 0.81797839
#>  [31,] 0.21223090 0.78776910
#>  [32,] 0.39496158 0.60503842
#>  [33,] 0.40761448 0.59238552
#>  [34,] 0.20787373 0.79212627
#>  [35,] 0.27950552 0.72049448
#>  [36,] 0.10895671 0.89104329
#>  [37,] 0.20564656 0.79435344
#>  [38,] 0.10895671 0.89104329
#>  [39,] 0.21514308 0.78485692
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.21514308 0.78485692
#>  [42,] 0.09677485 0.90322515
#>  [43,] 0.09668985 0.90331015
#>  [44,] 0.11109889 0.88890111
#>  [45,] 0.11109889 0.88890111
#>  [46,] 0.18281568 0.81718432
#>  [47,] 0.31242266 0.68757734
#>  [48,] 0.08182984 0.91817016
#>  [49,] 0.10895671 0.89104329
#>  [50,] 0.19078655 0.80921345
#>  [51,] 0.22005560 0.77994440
#>  [52,] 0.09668985 0.90331015
#>  [53,] 0.17851969 0.82148031
#>  [54,] 0.17851969 0.82148031
#>  [55,] 0.17851969 0.82148031
#>  [56,] 0.20287622 0.79712378
#>  [57,] 0.19346469 0.80653531
#>  [58,] 0.36133537 0.63866463
#>  [59,] 0.36133537 0.63866463
#>  [60,] 0.30508223 0.69491777
#>  [61,] 0.30332609 0.69667391
#>  [62,] 0.20778874 0.79221126
#>  [63,] 0.28891705 0.71108295
#>  [64,] 0.27581319 0.72418681
#>  [65,] 0.19078655 0.80921345
#>  [66,] 0.38301375 0.61698625
#>  [67,] 0.38111577 0.61888423
#>  [68,] 0.09668985 0.90331015
#>  [69,] 1.00000000 0.00000000
#>  [70,] 0.67667021 0.32332979
#>  [71,] 0.58165151 0.41834849
#>  [72,] 0.70404128 0.29595872
#>  [73,] 0.81797839 0.18202161
#>  [74,] 0.72418681 0.27581319
#>  [75,] 0.71369000 0.28631000
#>  [76,] 0.81797839 0.18202161
#>  [77,] 0.81797839 0.18202161
#>  [78,] 0.81376740 0.18623260
#>  [79,] 0.70687950 0.29312050
#>  [80,] 0.89205749 0.10794251
#>  [81,] 0.79212627 0.20787373
#>  [82,] 0.79186572 0.20813428
#>  [83,] 0.79186572 0.20813428
#>  [84,] 0.89205749 0.10794251
#>  [85,] 1.00000000 0.00000000
#>  [86,] 0.79893677 0.20106323
#>  [87,] 0.80500246 0.19499754
#>  [88,] 0.89104329 0.10895671
#>  [89,] 0.79426844 0.20573156
#>  [90,] 0.70902168 0.29097832
#>  [91,] 0.61233183 0.38766817
#>  [92,] 0.70902168 0.29097832
#>  [93,] 1.00000000 0.00000000
#>  [94,] 0.79703879 0.20296121
#>  [95,] 1.00000000 0.00000000
#>  [96,] 1.00000000 0.00000000
#>  [97,] 0.68593990 0.31406010
#>  [98,] 0.81797839 0.18202161
#>  [99,] 0.89980824 0.10019176
#> [100,] 0.89571161 0.10428839
#> [101,] 0.89980824 0.10019176
#> [102,] 0.69857189 0.30142811
#> [103,] 0.72784094 0.27215906
#> [104,] 0.78461272 0.21538728
#> [105,] 0.60329265 0.39670735
#> [106,] 0.60153651 0.39846349
#> [107,] 0.70404128 0.29595872
#> [108,] 0.71179202 0.28820798
#> [109,] 0.81797839 0.18202161
#> [110,] 0.60283531 0.39716469
#> [111,] 0.79703879 0.20296121
#> [112,] 0.79186572 0.20813428
#> [113,] 0.78870935 0.21129065
#> [114,] 0.90322515 0.09677485
#> [115,] 0.80707127 0.19292873
#> [116,] 0.81022764 0.18977236
#> [117,] 0.82148031 0.17851969
#> [118,] 0.82148031 0.17851969
#> [119,] 0.89980824 0.10019176
#> [120,] 0.79085152 0.20914848
#> [121,] 0.82139531 0.17860469
#> [122,] 0.68076683 0.31923317
#> [123,] 0.89980824 0.10019176
#> [124,] 1.00000000 0.00000000
#> [125,] 0.78587112 0.21412888
#> [126,] 0.89205749 0.10794251
#> [127,] 1.00000000 0.00000000
#> [128,] 1.00000000 0.00000000
#> [129,] 0.89205749 0.10794251
#> [130,] 0.79186572 0.20813428
#> [131,] 0.79186572 0.20813428
#> [132,] 1.00000000 0.00000000
#> [133,] 0.89980824 0.10019176
#> [134,] 0.79435344 0.20564656
#> [135,] 1.00000000 0.00000000
#> [136,] 1.00000000 0.00000000
#> [137,] 1.00000000 0.00000000
#> [138,] 0.71003588 0.28996412
#> [139,] 0.79902177 0.20097823
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  8.0028591  0.0000000  0.8472319 15.1721550  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.0079778  0.0000000  0.0000000  0.0000000  0.0000000  3.0283857  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  1.5289444  3.5047946  1.8575925  0.0000000  0.0000000  4.2584261  3.1723217 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  1.3674495  1.6963249  0.0000000  2.0555114 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  0.0000000  6.5410406  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  2.5346604  1.5453372  2.0777222  5.5345855  3.4263923  3.1783504 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  0.7675925  8.6990865  0.0000000  0.0000000  1.2415894  3.0222291  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  1.9470639  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  2.3964501 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  9.5879254 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 71 68 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V12         V9        V49         V1        V37        V45        V27 
#> 15.1721550  9.5879254  8.6990865  8.0028591  6.5410406  5.5345855  4.2584261 
#>        V23        V46        V47        V28        V20        V52        V42 
#>  3.5047946  3.4263923  3.1783504  3.1723217  3.0283857  3.0222291  2.5346604 
#>         V6        V44        V34        V54        V24        V32        V43 
#>  2.3964501  2.0777222  2.0555114  1.9470639  1.8575925  1.6963249  1.5453372 
#>        V22        V31        V51        V16        V11        V48        V10 
#>  1.5289444  1.3674495  1.2415894  1.0079778  0.8472319  0.7675925  0.0000000 
#>        V13        V14        V15        V17        V18        V19         V2 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V21        V25        V26        V29         V3        V30        V33 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V38        V39         V4        V40        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V5        V50        V53        V55        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913