Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 61 M (0.56115108 0.43884892)  
#>   2) V9>=0.10195 99 25 M (0.74747475 0.25252525)  
#>     4) V46>=0.0665 81 11 M (0.86419753 0.13580247)  
#>       8) V16< 0.6982 70  4 M (0.94285714 0.05714286) *
#>       9) V16>=0.6982 11  4 R (0.36363636 0.63636364) *
#>     5) V46< 0.0665 18  4 R (0.22222222 0.77777778) *
#>   3) V9< 0.10195 40  4 R (0.10000000 0.90000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V9>=0.11595 83 24 M (0.71084337 0.28915663)  
#>      4) V52>=0.0085 53  5 M (0.90566038 0.09433962) *
#>      5) V52< 0.0085 30 11 R (0.36666667 0.63333333)  
#>       10) V45>=0.18805 12  3 M (0.75000000 0.25000000) *
#>       11) V45< 0.18805 18  2 R (0.11111111 0.88888889) *
#>    3) V9< 0.11595 56 10 R (0.17857143 0.82142857)  
#>      6) V4>=0.0809 9  0 M (1.00000000 0.00000000) *
#>      7) V4< 0.0809 47  1 R (0.02127660 0.97872340) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 58 M (0.58273381 0.41726619)  
#>   2) V13>=0.2279 90 23 M (0.74444444 0.25555556)  
#>     4) V23>=0.3575 77 12 M (0.84415584 0.15584416)  
#>       8) V8>=0.08225 63  4 M (0.93650794 0.06349206) *
#>       9) V8< 0.08225 14  6 R (0.42857143 0.57142857) *
#>     5) V23< 0.3575 13  2 R (0.15384615 0.84615385) *
#>   3) V13< 0.2279 49 14 R (0.28571429 0.71428571)  
#>     6) V25< 0.55335 19  7 M (0.63157895 0.36842105) *
#>     7) V25>=0.55335 30  2 R (0.06666667 0.93333333) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.5395683 0.4604317)  
#>    2) V4>=0.0606 45  5 M (0.8888889 0.1111111) *
#>    3) V4< 0.0606 94 35 R (0.3723404 0.6276596)  
#>      6) V28>=0.9273 22  3 M (0.8636364 0.1363636) *
#>      7) V28< 0.9273 72 16 R (0.2222222 0.7777778)  
#>       14) V10>=0.3119 7  0 M (1.0000000 0.0000000) *
#>       15) V10< 0.3119 65  9 R (0.1384615 0.8615385)  
#>         30) V51>=0.01365 25  9 R (0.3600000 0.6400000)  
#>           60) V12>=0.18345 8  1 M (0.8750000 0.1250000) *
#>           61) V12< 0.18345 17  2 R (0.1176471 0.8823529) *
#>         31) V51< 0.01365 40  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V37< 0.1859 40  6 M (0.85000000 0.15000000)  
#>      4) V48>=0.04555 32  0 M (1.00000000 0.00000000) *
#>      5) V48< 0.04555 8  2 R (0.25000000 0.75000000) *
#>    3) V37>=0.1859 99 35 R (0.35353535 0.64646465)  
#>      6) V45>=0.2797 20  3 M (0.85000000 0.15000000) *
#>      7) V45< 0.2797 79 18 R (0.22784810 0.77215190)  
#>       14) V54>=0.0225 8  0 M (1.00000000 0.00000000) *
#>       15) V54< 0.0225 71 10 R (0.14084507 0.85915493)  
#>         30) V12>=0.22525 28  9 R (0.32142857 0.67857143)  
#>           60) V38>=0.20545 15  6 M (0.60000000 0.40000000) *
#>           61) V38< 0.20545 13  0 R (0.00000000 1.00000000) *
#>         31) V12< 0.22525 43  1 R (0.02325581 0.97674419) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V21>=0.7883 49  7 M (0.85714286 0.14285714)  
#>      4) V31< 0.4599 41  2 M (0.95121951 0.04878049) *
#>      5) V31>=0.4599 8  3 R (0.37500000 0.62500000) *
#>    3) V21< 0.7883 90 30 R (0.33333333 0.66666667)  
#>      6) V27>=0.90335 23  3 M (0.86956522 0.13043478) *
#>      7) V27< 0.90335 67 10 R (0.14925373 0.85074627)  
#>       14) V36< 0.40615 25  9 R (0.36000000 0.64000000)  
#>         28) V48>=0.05605 8  1 M (0.87500000 0.12500000) *
#>         29) V48< 0.05605 17  2 R (0.11764706 0.88235294) *
#>       15) V36>=0.40615 42  1 R (0.02380952 0.97619048) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 50 R (0.35971223 0.64028777)  
#>    2) V11>=0.19685 80 37 M (0.53750000 0.46250000)  
#>      4) V2< 0.03165 34  4 M (0.88235294 0.11764706)  
#>        8) V8< 0.17035 27  0 M (1.00000000 0.00000000) *
#>        9) V8>=0.17035 7  3 R (0.42857143 0.57142857) *
#>      5) V2>=0.03165 46 13 R (0.28260870 0.71739130)  
#>       10) V37< 0.3766 15  4 M (0.73333333 0.26666667) *
#>       11) V37>=0.3766 31  2 R (0.06451613 0.93548387) *
#>    3) V11< 0.19685 59  7 R (0.11864407 0.88135593) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V20>=0.8328 51 15 M (0.70588235 0.29411765)  
#>      4) V25< 0.66 31  1 M (0.96774194 0.03225806) *
#>      5) V25>=0.66 20  6 R (0.30000000 0.70000000)  
#>       10) V13>=0.268 8  2 M (0.75000000 0.25000000) *
#>       11) V13< 0.268 12  0 R (0.00000000 1.00000000) *
#>    3) V20< 0.8328 88 28 R (0.31818182 0.68181818)  
#>      6) V29>=0.81315 27  7 M (0.74074074 0.25925926)  
#>       12) V41>=0.1843 19  1 M (0.94736842 0.05263158) *
#>       13) V41< 0.1843 8  2 R (0.25000000 0.75000000) *
#>      7) V29< 0.81315 61  8 R (0.13114754 0.86885246) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V48>=0.07485 78 16 M (0.79487179 0.20512821)  
#>      4) V37< 0.84395 70  8 M (0.88571429 0.11428571)  
#>        8) V15< 0.55275 61  2 M (0.96721311 0.03278689) *
#>        9) V15>=0.55275 9  3 R (0.33333333 0.66666667) *
#>      5) V37>=0.84395 8  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.07485 61 20 R (0.32786885 0.67213115)  
#>      6) V12>=0.4502 9  0 M (1.00000000 0.00000000) *
#>      7) V12< 0.4502 52 11 R (0.21153846 0.78846154)  
#>       14) V4>=0.0614 11  4 M (0.63636364 0.36363636) *
#>       15) V4< 0.0614 41  4 R (0.09756098 0.90243902) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V40< 0.6328 112 41 M (0.63392857 0.36607143)  
#>      4) V50>=0.01885 66 10 M (0.84848485 0.15151515)  
#>        8) V58>=0.00325 59  5 M (0.91525424 0.08474576) *
#>        9) V58< 0.00325 7  2 R (0.28571429 0.71428571) *
#>      5) V50< 0.01885 46 15 R (0.32608696 0.67391304)  
#>       10) V28>=0.958 9  1 M (0.88888889 0.11111111) *
#>       11) V28< 0.958 37  7 R (0.18918919 0.81081081)  
#>         22) V19>=0.80455 8  2 M (0.75000000 0.25000000) *
#>         23) V19< 0.80455 29  1 R (0.03448276 0.96551724) *
#>    3) V40>=0.6328 27  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.7738443 0.5952754 0.8024807 0.7029829 0.8008872 0.5836285
#>  [8] 0.7698718 0.7215027 0.7514911
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.9581691 4.2551263
#>   [2,] 2.1192587 5.0940368
#>   [3,] 3.6231287 3.5901667
#>   [4,] 2.8522300 4.3610655
#>   [5,] 1.2982582 5.9150373
#>   [6,] 2.7717867 4.4415088
#>   [7,] 2.2637093 4.9495862
#>   [8,] 1.5523783 5.6609172
#>   [9,] 1.3977560 5.8155395
#>  [10,] 0.7113310 6.5019645
#>  [11,] 2.9382972 4.2749982
#>  [12,] 2.2653028 4.9479927
#>  [13,] 0.8008872 6.4124083
#>  [14,] 2.1881582 5.0251373
#>  [15,] 1.5054635 5.7078320
#>  [16,] 2.1192587 5.0940368
#>  [17,] 1.3167781 5.8965174
#>  [18,] 0.7698718 6.4434237
#>  [19,] 1.5539718 5.6593237
#>  [20,] 2.1076119 5.1056836
#>  [21,] 0.7698718 6.4434237
#>  [22,] 0.5836285 6.6296670
#>  [23,] 1.5539718 5.6593237
#>  [24,] 1.3691197 5.8441758
#>  [25,] 0.0000000 7.2132955
#>  [26,] 0.0000000 7.2132955
#>  [27,] 0.0000000 7.2132955
#>  [28,] 2.1206108 5.0926847
#>  [29,] 1.5213629 5.6919326
#>  [30,] 2.0197609 5.1935346
#>  [31,] 2.1476536 5.0656419
#>  [32,] 3.0987149 4.1145806
#>  [33,] 2.7712520 4.4420435
#>  [34,] 2.0197609 5.1935346
#>  [35,] 0.7029829 6.5103126
#>  [36,] 0.0000000 7.2132955
#>  [37,] 0.8008872 6.4124083
#>  [38,] 1.5223899 5.6909056
#>  [39,] 1.5523783 5.6609172
#>  [40,] 2.3146988 4.8985967
#>  [41,] 2.3248705 4.8884249
#>  [42,] 0.8008872 6.4124083
#>  [43,] 0.8008872 6.4124083
#>  [44,] 2.3548589 4.8584366
#>  [45,] 0.5952754 6.6180201
#>  [46,] 1.3167781 5.8965174
#>  [47,] 2.0866499 5.1266456
#>  [48,] 0.7029829 6.5103126
#>  [49,] 1.4143139 5.7989816
#>  [50,] 0.7113310 6.5019645
#>  [51,] 0.0000000 7.2132955
#>  [52,] 0.8008872 6.4124083
#>  [53,] 1.5523783 5.6609172
#>  [54,] 2.3262226 4.8870729
#>  [55,] 1.4768272 5.7364683
#>  [56,] 2.2553611 4.9579344
#>  [57,] 0.0000000 7.2132955
#>  [58,] 1.3691197 5.8441758
#>  [59,] 1.9487756 5.2645198
#>  [60,] 2.8079742 4.4053212
#>  [61,] 2.1716004 5.0416951
#>  [62,] 2.2366664 4.9766290
#>  [63,] 1.5138117 5.6994838
#>  [64,] 1.3574728 5.8558226
#>  [65,] 1.9487756 5.2645198
#>  [66,] 5.0589080 2.1543875
#>  [67,] 5.0472612 2.1660343
#>  [68,] 5.6919326 1.5213629
#>  [69,] 4.9272330 2.2860625
#>  [70,] 4.5205800 2.6927155
#>  [71,] 6.5019645 0.7113310
#>  [72,] 4.6200778 2.5932177
#>  [73,] 5.2153531 1.9979424
#>  [74,] 5.9183360 1.2949595
#>  [75,] 6.0343916 1.1789039
#>  [76,] 5.6893121 1.5239834
#>  [77,] 7.2132955 0.0000000
#>  [78,] 5.7320927 1.4812028
#>  [79,] 4.9210338 2.2922617
#>  [80,] 4.9806016 2.2326939
#>  [81,] 5.1343199 2.0789756
#>  [82,] 4.9665965 2.2466990
#>  [83,] 6.4434237 0.7698718
#>  [84,] 4.9889497 2.2243457
#>  [85,] 6.4917928 0.7215027
#>  [86,] 5.7804618 1.4328337
#>  [87,] 4.4229889 2.7903066
#>  [88,] 5.7320927 1.4812028
#>  [89,] 5.7320927 1.4812028
#>  [90,] 6.6296670 0.5836285
#>  [91,] 5.8597952 1.3535003
#>  [92,] 6.0343916 1.1789039
#>  [93,] 5.0772887 2.1360068
#>  [94,] 4.5130287 2.7002667
#>  [95,] 6.6296670 0.5836285
#>  [96,] 7.2132955 0.0000000
#>  [97,] 5.0472612 2.1660343
#>  [98,] 5.0589080 2.1543875
#>  [99,] 6.4434237 0.7698718
#> [100,] 7.2132955 0.0000000
#> [101,] 5.2645198 1.9487756
#> [102,] 4.9965010 2.2167945
#> [103,] 5.8271863 1.3861092
#> [104,] 4.4027008 2.8105947
#> [105,] 5.0262992 2.1869963
#> [106,] 7.2132955 0.0000000
#> [107,] 5.6609172 1.5523783
#> [108,] 6.0343916 1.1789039
#> [109,] 5.2335045 1.9797910
#> [110,] 6.0343916 1.1789039
#> [111,] 5.0966572 2.1166382
#> [112,] 5.0589080 2.1543875
#> [113,] 5.8271863 1.3861092
#> [114,] 5.1444916 2.0688039
#> [115,] 5.8965174 1.3167781
#> [116,] 6.6180201 0.5952754
#> [117,] 5.8597952 1.3535003
#> [118,] 5.2829005 1.9303949
#> [119,] 4.9889497 2.2243457
#> [120,] 5.6919326 1.5213629
#> [121,] 5.1083041 2.1049914
#> [122,] 6.0343916 1.1789039
#> [123,] 5.6919326 1.5213629
#> [124,] 6.4618044 0.7514911
#> [125,] 6.4124083 0.8008872
#> [126,] 6.4434237 0.7698718
#> [127,] 6.5103126 0.7029829
#> [128,] 7.2132955 0.0000000
#> [129,] 7.2132955 0.0000000
#> [130,] 6.4394511 0.7738443
#> [131,] 5.8441758 1.3691197
#> [132,] 5.6695793 1.5437161
#> [133,] 5.6919326 1.5213629
#> [134,] 5.6919326 1.5213629
#> [135,] 6.4434237 0.7698718
#> [136,] 5.7219210 1.4913745
#> [137,] 4.9582483 2.2550471
#> [138,] 5.8481483 1.3651471
#> [139,] 5.8481483 1.3651471
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.41009954 0.58990046
#>   [2,] 0.29379896 0.70620104
#>   [3,] 0.50228481 0.49771519
#>   [4,] 0.39541288 0.60458712
#>   [5,] 0.17998129 0.82001871
#>   [6,] 0.38426080 0.61573920
#>   [7,] 0.31382456 0.68617544
#>   [8,] 0.21521068 0.78478932
#>   [9,] 0.19377496 0.80622504
#>  [10,] 0.09861387 0.90138613
#>  [11,] 0.40734464 0.59265536
#>  [12,] 0.31404547 0.68595453
#>  [13,] 0.11102930 0.88897070
#>  [14,] 0.30335070 0.69664930
#>  [15,] 0.20870676 0.79129324
#>  [16,] 0.29379896 0.70620104
#>  [17,] 0.18254875 0.81745125
#>  [18,] 0.10672955 0.89327045
#>  [19,] 0.21543160 0.78456840
#>  [20,] 0.29218433 0.70781567
#>  [21,] 0.10672955 0.89327045
#>  [22,] 0.08091011 0.91908989
#>  [23,] 0.21543160 0.78456840
#>  [24,] 0.18980502 0.81019498
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.29398640 0.70601360
#>  [29,] 0.21091093 0.78908907
#>  [30,] 0.28000529 0.71999471
#>  [31,] 0.29773543 0.70226457
#>  [32,] 0.42958380 0.57041620
#>  [33,] 0.38418668 0.61581332
#>  [34,] 0.28000529 0.71999471
#>  [35,] 0.09745655 0.90254345
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.11102930 0.88897070
#>  [38,] 0.21105331 0.78894669
#>  [39,] 0.21521068 0.78478932
#>  [40,] 0.32089339 0.67910661
#>  [41,] 0.32230352 0.67769648
#>  [42,] 0.11102930 0.88897070
#>  [43,] 0.11102930 0.88897070
#>  [44,] 0.32646090 0.67353910
#>  [45,] 0.08252474 0.91747526
#>  [46,] 0.18254875 0.81745125
#>  [47,] 0.28927830 0.71072170
#>  [48,] 0.09745655 0.90254345
#>  [49,] 0.19607042 0.80392958
#>  [50,] 0.09861387 0.90138613
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.11102930 0.88897070
#>  [53,] 0.21521068 0.78478932
#>  [54,] 0.32249096 0.67750904
#>  [55,] 0.20473682 0.79526318
#>  [56,] 0.31266723 0.68733277
#>  [57,] 0.00000000 1.00000000
#>  [58,] 0.18980502 0.81019498
#>  [59,] 0.27016440 0.72983560
#>  [60,] 0.38927759 0.61072241
#>  [61,] 0.30105523 0.69894477
#>  [62,] 0.31007553 0.68992447
#>  [63,] 0.20986409 0.79013591
#>  [64,] 0.18819038 0.81180962
#>  [65,] 0.27016440 0.72983560
#>  [66,] 0.70133104 0.29866896
#>  [67,] 0.69971640 0.30028360
#>  [68,] 0.78908907 0.21091093
#>  [69,] 0.68307655 0.31692345
#>  [70,] 0.62670107 0.37329893
#>  [71,] 0.90138613 0.09861387
#>  [72,] 0.64049473 0.35950527
#>  [73,] 0.72301948 0.27698052
#>  [74,] 0.82047602 0.17952398
#>  [75,] 0.83656515 0.16343485
#>  [76,] 0.78872578 0.21127422
#>  [77,] 1.00000000 0.00000000
#>  [78,] 0.79465657 0.20534343
#>  [79,] 0.68221714 0.31778286
#>  [80,] 0.69047519 0.30952481
#>  [81,] 0.71178561 0.28821439
#>  [82,] 0.68853362 0.31146638
#>  [83,] 0.89327045 0.10672955
#>  [84,] 0.69163252 0.30836748
#>  [85,] 0.89997599 0.10002401
#>  [86,] 0.80136212 0.19863788
#>  [87,] 0.61317174 0.38682826
#>  [88,] 0.79465657 0.20534343
#>  [89,] 0.79465657 0.20534343
#>  [90,] 0.91908989 0.08091011
#>  [91,] 0.81236034 0.18763966
#>  [92,] 0.83656515 0.16343485
#>  [93,] 0.70387921 0.29612079
#>  [94,] 0.62565422 0.37434578
#>  [95,] 0.91908989 0.08091011
#>  [96,] 1.00000000 0.00000000
#>  [97,] 0.69971640 0.30028360
#>  [98,] 0.70133104 0.29866896
#>  [99,] 0.89327045 0.10672955
#> [100,] 1.00000000 0.00000000
#> [101,] 0.72983560 0.27016440
#> [102,] 0.69267937 0.30732063
#> [103,] 0.80783968 0.19216032
#> [104,] 0.61035913 0.38964087
#> [105,] 0.69681038 0.30318962
#> [106,] 1.00000000 0.00000000
#> [107,] 0.78478932 0.21521068
#> [108,] 0.83656515 0.16343485
#> [109,] 0.72553585 0.27446415
#> [110,] 0.83656515 0.16343485
#> [111,] 0.70656432 0.29343568
#> [112,] 0.70133104 0.29866896
#> [113,] 0.80783968 0.19216032
#> [114,] 0.71319574 0.28680426
#> [115,] 0.81745125 0.18254875
#> [116,] 0.91747526 0.08252474
#> [117,] 0.81236034 0.18763966
#> [118,] 0.73238377 0.26761623
#> [119,] 0.69163252 0.30836748
#> [120,] 0.78908907 0.21091093
#> [121,] 0.70817896 0.29182104
#> [122,] 0.83656515 0.16343485
#> [123,] 0.78908907 0.21091093
#> [124,] 0.89581862 0.10418138
#> [125,] 0.88897070 0.11102930
#> [126,] 0.89327045 0.10672955
#> [127,] 0.90254345 0.09745655
#> [128,] 1.00000000 0.00000000
#> [129,] 1.00000000 0.00000000
#> [130,] 0.89271972 0.10728028
#> [131,] 0.81019498 0.18980502
#> [132,] 0.78599017 0.21400983
#> [133,] 0.78908907 0.21091093
#> [134,] 0.78908907 0.21091093
#> [135,] 0.89327045 0.10672955
#> [136,] 0.79324644 0.20675356
#> [137,] 0.68737630 0.31262370
#> [138,] 0.81074571 0.18925429
#> [139,] 0.81074571 0.18925429
#> 
#> $class
#>   [1] "R" "R" "M" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 2.2330216 2.0626386 4.1556887 3.5908280 0.0000000 1.3488268 1.3458474 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 1.4311344 2.4347005 2.2165594 4.1359189 0.0000000 1.8716624 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 3.7865973 0.0000000 4.2208727 4.8804431 3.1764328 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 1.0559786 0.0000000 0.0000000 0.0000000 0.0000000 0.8415632 6.9047261 1.0455720 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 8.2649334 3.8976388 1.2504113 0.0000000 0.0000000 0.0000000 3.9254301 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 2.5613022 0.0000000 6.1791809 0.0000000 0.0000000 3.2979231 0.9491970 2.5548779 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 2.2132938 0.0000000 0.0000000 0.0000000 1.1056220 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 1.6720795 9.3890975 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V9        V4       V37       V48       V28       V27       V12       V21 
#> 9.3890975 8.2649334 6.9047261 6.1791809 4.8804431 4.2208727 4.1556887 4.1359189 
#>       V45       V40       V25       V13       V50       V29       V46       V52 
#> 3.9254301 3.8976388 3.7865973 3.5908280 3.2979231 3.1764328 2.5613022 2.5548779 
#>        V2       V10       V20       V54       V11       V23        V8       V19 
#> 2.4347005 2.2330216 2.2165594 2.2132938 2.0626386 1.8716624 1.6720795 1.4311344 
#>       V15       V16       V41       V58       V31       V38       V51       V36 
#> 1.3488268 1.3458474 1.2504113 1.1056220 1.0559786 1.0455720 0.9491970 0.8415632 
#>        V1       V14       V17       V18       V22       V24       V26        V3 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V30       V32       V33       V34       V35       V39       V42       V43 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V44       V47       V49        V5       V53       V55       V56       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V59        V6       V60        V7 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058