Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V11>=0.16895 86 20 M (0.76744186 0.23255814)  
#>      4) V37< 0.38475 53  2 M (0.96226415 0.03773585) *
#>      5) V37>=0.38475 33 15 R (0.45454545 0.54545455)  
#>       10) V17< 0.39985 15  1 M (0.93333333 0.06666667) *
#>       11) V17>=0.39985 18  1 R (0.05555556 0.94444444) *
#>    3) V11< 0.16895 53  7 R (0.13207547 0.86792453)  
#>      6) V5>=0.0923 7  1 M (0.85714286 0.14285714) *
#>      7) V5< 0.0923 46  1 R (0.02173913 0.97826087) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V47>=0.04705 109 34 M (0.68807339 0.31192661)  
#>      4) V9>=0.10655 88 18 M (0.79545455 0.20454545)  
#>        8) V15< 0.56165 71  5 M (0.92957746 0.07042254) *
#>        9) V15>=0.56165 17  4 R (0.23529412 0.76470588) *
#>      5) V9< 0.10655 21  5 R (0.23809524 0.76190476)  
#>       10) V1< 0.01445 7  2 M (0.71428571 0.28571429) *
#>       11) V1>=0.01445 14  0 R (0.00000000 1.00000000) *
#>    3) V47< 0.04705 30  1 R (0.03333333 0.96666667) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V49>=0.0376 67 16 M (0.76119403 0.23880597)  
#>      4) V9>=0.1117 50  4 M (0.92000000 0.08000000) *
#>      5) V9< 0.1117 17  5 R (0.29411765 0.70588235) *
#>    3) V49< 0.0376 72 16 R (0.22222222 0.77777778)  
#>      6) V44>=0.2489 13  3 M (0.76923077 0.23076923) *
#>      7) V44< 0.2489 59  6 R (0.10169492 0.89830508)  
#>       14) V27< 0.50185 7  3 M (0.57142857 0.42857143) *
#>       15) V27>=0.50185 52  2 R (0.03846154 0.96153846) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V35< 0.22025 40  5 M (0.87500000 0.12500000) *
#>    3) V35>=0.22025 99 36 R (0.36363636 0.63636364)  
#>      6) V54>=0.01755 22  1 M (0.95454545 0.04545455) *
#>      7) V54< 0.01755 77 15 R (0.19480519 0.80519481)  
#>       14) V45>=0.32855 10  3 M (0.70000000 0.30000000) *
#>       15) V45< 0.32855 67  8 R (0.11940299 0.88059701)  
#>         30) V21>=0.68165 26  8 R (0.30769231 0.69230769)  
#>           60) V1< 0.03075 10  2 M (0.80000000 0.20000000) *
#>           61) V1>=0.03075 16  0 R (0.00000000 1.00000000) *
#>         31) V21< 0.68165 41  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 52 R (0.37410072 0.62589928)  
#>    2) V49>=0.04525 61 20 M (0.67213115 0.32786885)  
#>      4) V28>=0.74365 28  0 M (1.00000000 0.00000000) *
#>      5) V28< 0.74365 33 13 R (0.39393939 0.60606061)  
#>       10) V58>=0.0069 16  3 M (0.81250000 0.18750000) *
#>       11) V58< 0.0069 17  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.04525 78 11 R (0.14102564 0.85897436)  
#>      6) V45>=0.25695 7  1 M (0.85714286 0.14285714) *
#>      7) V45< 0.25695 71  5 R (0.07042254 0.92957746)  
#>       14) V48< 0.0227 7  3 M (0.57142857 0.42857143) *
#>       15) V48>=0.0227 64  1 R (0.01562500 0.98437500) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V51>=0.01555 57 13 M (0.77192982 0.22807018)  
#>      4) V40< 0.56475 49  5 M (0.89795918 0.10204082)  
#>        8) V48>=0.06365 41  0 M (1.00000000 0.00000000) *
#>        9) V48< 0.06365 8  3 R (0.37500000 0.62500000) *
#>      5) V40>=0.56475 8  0 R (0.00000000 1.00000000) *
#>    3) V51< 0.01555 82 20 R (0.24390244 0.75609756)  
#>      6) V22>=0.9281 12  1 M (0.91666667 0.08333333) *
#>      7) V22< 0.9281 70  9 R (0.12857143 0.87142857)  
#>       14) V28>=0.9767 7  0 M (1.00000000 0.00000000) *
#>       15) V28< 0.9767 63  2 R (0.03174603 0.96825397) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 68 R (0.48920863 0.51079137)  
#>   2) V36< 0.4434 87 28 M (0.67816092 0.32183908)  
#>     4) V47>=0.0582 70 14 M (0.80000000 0.20000000)  
#>       8) V5>=0.02095 62  6 M (0.90322581 0.09677419) *
#>       9) V5< 0.02095 8  0 R (0.00000000 1.00000000) *
#>     5) V47< 0.0582 17  3 R (0.17647059 0.82352941) *
#>   3) V36>=0.4434 52  9 R (0.17307692 0.82692308)  
#>     6) V45>=0.3857 7  0 M (1.00000000 0.00000000) *
#>     7) V45< 0.3857 45  2 R (0.04444444 0.95555556) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 56 R (0.40287770 0.59712230)  
#>    2) V54>=0.0207 15  0 M (1.00000000 0.00000000) *
#>    3) V54< 0.0207 124 41 R (0.33064516 0.66935484)  
#>      6) V9>=0.13685 66 31 M (0.53030303 0.46969697)  
#>       12) V36< 0.41575 43 11 M (0.74418605 0.25581395)  
#>         24) V55< 0.00915 27  0 M (1.00000000 0.00000000) *
#>         25) V55>=0.00915 16  5 R (0.31250000 0.68750000) *
#>       13) V36>=0.41575 23  3 R (0.13043478 0.86956522) *
#>      7) V9< 0.13685 58  6 R (0.10344828 0.89655172)  
#>       14) V28>=0.9356 7  2 M (0.71428571 0.28571429) *
#>       15) V28< 0.9356 51  1 R (0.01960784 0.98039216) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V43>=0.09515 112 46 M (0.58928571 0.41071429)  
#>      4) V21>=0.55655 73 19 M (0.73972603 0.26027397)  
#>        8) V51>=0.0108 41  1 M (0.97560976 0.02439024) *
#>        9) V51< 0.0108 32 14 R (0.43750000 0.56250000)  
#>         18) V55< 0.0039 8  0 M (1.00000000 0.00000000) *
#>         19) V55>=0.0039 24  6 R (0.25000000 0.75000000)  
#>           38) V42>=0.3086 7  1 M (0.85714286 0.14285714) *
#>           39) V42< 0.3086 17  0 R (0.00000000 1.00000000) *
#>      5) V21< 0.55655 39 12 R (0.30769231 0.69230769)  
#>       10) V47>=0.1504 15  3 M (0.80000000 0.20000000) *
#>       11) V47< 0.1504 24  0 R (0.00000000 1.00000000) *
#>    3) V43< 0.09515 27  2 R (0.07407407 0.92592593) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.5179856 0.4820144)  
#>    2) V12>=0.20075 71 14 M (0.8028169 0.1971831)  
#>      4) V3>=0.0237 56  6 M (0.8928571 0.1071429) *
#>      5) V3< 0.0237 15  7 R (0.4666667 0.5333333) *
#>    3) V12< 0.20075 68 15 R (0.2205882 0.7794118)  
#>      6) V31< 0.4623 23  8 M (0.6521739 0.3478261)  
#>       12) V7>=0.13805 14  0 M (1.0000000 0.0000000) *
#>       13) V7< 0.13805 9  1 R (0.1111111 0.8888889) *
#>      7) V31>=0.4623 45  0 R (0.0000000 1.0000000) *
#> 
#> 
#> $weights
#>  [1] 0.8630811 0.7473342 0.5969304 0.7216073 0.5943095 0.7678189 0.8142241
#>  [8] 0.8304135 0.8130081 0.6232537
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.1590850 5.2128957
#>   [2,] 2.2819495 5.0900312
#>   [3,] 1.9391704 5.4328103
#>   [4,] 2.3005588 5.0714219
#>   [5,] 2.1301408 5.2418399
#>   [6,] 1.5346153 5.8373654
#>   [7,] 0.0000000 7.3719807
#>   [8,] 2.6891256 4.6828551
#>   [9,] 1.5615583 5.8104224
#>  [10,] 2.4234234 4.9485573
#>  [11,] 1.6309000 5.7410807
#>  [12,] 3.0209501 4.3510307
#>  [13,] 2.5618278 4.8101529
#>  [14,] 2.5618278 4.8101529
#>  [15,] 3.6904831 3.6814976
#>  [16,] 2.1751365 5.1968442
#>  [17,] 1.1912399 6.1807408
#>  [18,] 1.4374777 5.9345030
#>  [19,] 2.0658719 5.3061088
#>  [20,] 2.1816188 5.1903619
#>  [21,] 1.4863347 5.8856460
#>  [22,] 0.8630811 6.5088996
#>  [23,] 1.6104153 5.7615654
#>  [24,] 0.0000000 7.3719807
#>  [25,] 0.0000000 7.3719807
#>  [26,] 0.0000000 7.3719807
#>  [27,] 1.5151532 5.8568275
#>  [28,] 0.5943095 6.7776712
#>  [29,] 0.0000000 7.3719807
#>  [30,] 2.3455666 5.0264141
#>  [31,] 2.7521785 4.6198023
#>  [32,] 2.1572727 5.2147080
#>  [33,] 0.5943095 6.7776712
#>  [34,] 0.7216073 6.6503735
#>  [35,] 0.7216073 6.6503735
#>  [36,] 2.4151018 4.9568789
#>  [37,] 0.0000000 7.3719807
#>  [38,] 0.5943095 6.7776712
#>  [39,] 0.5943095 6.7776712
#>  [40,] 0.0000000 7.3719807
#>  [41,] 1.5982324 5.7737483
#>  [42,] 1.3442647 6.0277160
#>  [43,] 0.5943095 6.7776712
#>  [44,] 0.6232537 6.7487271
#>  [45,] 1.9128472 5.4591335
#>  [46,] 0.0000000 7.3719807
#>  [47,] 0.0000000 7.3719807
#>  [48,] 0.7216073 6.6503735
#>  [49,] 0.7216073 6.6503735
#>  [50,] 2.3198397 5.0521411
#>  [51,] 3.0509845 4.3209962
#>  [52,] 2.1590850 5.2128957
#>  [53,] 1.4689415 5.9030392
#>  [54,] 2.8962732 4.4757075
#>  [55,] 2.8974893 4.4744915
#>  [56,] 2.0331922 5.3387886
#>  [57,] 2.0042480 5.3677327
#>  [58,] 2.1777574 5.1942233
#>  [59,] 2.0806442 5.2913365
#>  [60,] 1.4863347 5.8856460
#>  [61,] 1.5615583 5.8104224
#>  [62,] 1.9128472 5.4591335
#>  [63,] 5.0057359 2.3662448
#>  [64,] 4.9607403 2.4112405
#>  [65,] 4.2596177 3.1123630
#>  [66,] 5.0426035 2.3293773
#>  [67,] 4.3163577 3.0556230
#>  [68,] 5.7899377 1.5820430
#>  [69,] 5.1973026 2.1746781
#>  [70,] 4.5766699 2.7953109
#>  [71,] 5.1003661 2.2716147
#>  [72,] 6.1807408 1.1912399
#>  [73,] 6.7776712 0.5943095
#>  [74,] 6.7487271 0.6232537
#>  [75,] 5.7737483 1.5982324
#>  [76,] 6.1544176 1.2175632
#>  [77,] 6.6041618 0.7678189
#>  [78,] 7.3719807 0.0000000
#>  [79,] 5.3677327 2.0042480
#>  [80,] 4.9812250 2.3907558
#>  [81,] 5.7942330 1.5777477
#>  [82,] 5.2625180 2.1094627
#>  [83,] 5.7737483 1.5982324
#>  [84,] 6.7487271 0.6232537
#>  [85,] 5.8568275 1.5151532
#>  [86,] 6.5577566 0.8142241
#>  [87,] 5.7737483 1.5982324
#>  [88,] 6.5577566 0.8142241
#>  [89,] 6.0277160 1.3442647
#>  [90,] 6.5577566 0.8142241
#>  [91,] 6.5088996 0.8630811
#>  [92,] 7.3719807 0.0000000
#>  [93,] 6.5415673 0.8304135
#>  [94,] 5.8856460 1.4863347
#>  [95,] 6.5415673 0.8304135
#>  [96,] 6.7487271 0.6232537
#>  [97,] 5.1053056 2.2666752
#>  [98,] 5.1504391 2.2215416
#>  [99,] 5.8856460 1.4863347
#> [100,] 5.9809081 1.3910726
#> [101,] 5.3176597 2.0543210
#> [102,] 5.9145901 1.4573906
#> [103,] 5.1504947 2.2214861
#> [104,] 5.7737483 1.5982324
#> [105,] 5.7737483 1.5982324
#> [106,] 5.1504947 2.2214861
#> [107,] 6.0560640 1.3159168
#> [108,] 5.9809081 1.3910726
#> [109,] 5.1040895 2.2678912
#> [110,] 5.2887155 2.0832652
#> [111,] 5.8361494 1.5358313
#> [112,] 5.1504947 2.2214861
#> [113,] 5.4129218 1.9590589
#> [114,] 5.1942233 2.1777574
#> [115,] 6.5415673 0.8304135
#> [116,] 5.9809081 1.3910726
#> [117,] 5.0426035 2.3293773
#> [118,] 5.1768179 2.1951629
#> [119,] 5.8825545 1.4894262
#> [120,] 7.3719807 0.0000000
#> [121,] 6.5415673 0.8304135
#> [122,] 6.5415673 0.8304135
#> [123,] 6.5415673 0.8304135
#> [124,] 6.5415673 0.8304135
#> [125,] 6.6041618 0.7678189
#> [126,] 5.9809081 1.3910726
#> [127,] 6.0277160 1.3442647
#> [128,] 5.5574871 1.8144936
#> [129,] 6.7487271 0.6232537
#> [130,] 6.5589727 0.8130081
#> [131,] 7.3719807 0.0000000
#> [132,] 7.3719807 0.0000000
#> [133,] 5.3665167 2.0054640
#> [134,] 6.6041618 0.7678189
#> [135,] 6.5589727 0.8130081
#> [136,] 5.2147080 2.1572727
#> [137,] 5.9357190 1.4362617
#> [138,] 5.7899377 1.5820430
#> [139,] 5.3839777 1.9880030
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.29287719 0.70712281
#>   [2,] 0.30954361 0.69045639
#>   [3,] 0.26304605 0.73695395
#>   [4,] 0.31206794 0.68793206
#>   [5,] 0.28895095 0.71104905
#>   [6,] 0.20816865 0.79183135
#>   [7,] 0.00000000 1.00000000
#>   [8,] 0.36477654 0.63522346
#>   [9,] 0.21182344 0.78817656
#>  [10,] 0.32873436 0.67126564
#>  [11,] 0.22122956 0.77877044
#>  [12,] 0.40978811 0.59021189
#>  [13,] 0.34750875 0.65249125
#>  [14,] 0.34750875 0.65249125
#>  [15,] 0.50060944 0.49939056
#>  [16,] 0.29505456 0.70494544
#>  [17,] 0.16159021 0.83840979
#>  [18,] 0.19499206 0.80500794
#>  [19,] 0.28023295 0.71976705
#>  [20,] 0.29593387 0.70406613
#>  [21,] 0.20161946 0.79838054
#>  [22,] 0.11707588 0.88292412
#>  [23,] 0.21845083 0.78154917
#>  [24,] 0.00000000 1.00000000
#>  [25,] 0.00000000 1.00000000
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.20552864 0.79447136
#>  [28,] 0.08061734 0.91938266
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.31817319 0.68182681
#>  [31,] 0.37332958 0.62667042
#>  [32,] 0.29263136 0.70736864
#>  [33,] 0.08061734 0.91938266
#>  [34,] 0.09788512 0.90211488
#>  [35,] 0.09788512 0.90211488
#>  [36,] 0.32760555 0.67239445
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.08061734 0.91938266
#>  [39,] 0.08061734 0.91938266
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.21679823 0.78320177
#>  [42,] 0.18234783 0.81765217
#>  [43,] 0.08061734 0.91938266
#>  [44,] 0.08454358 0.91545642
#>  [45,] 0.25947534 0.74052466
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.09788512 0.90211488
#>  [49,] 0.09788512 0.90211488
#>  [50,] 0.31468336 0.68531664
#>  [51,] 0.41386225 0.58613775
#>  [52,] 0.29287719 0.70712281
#>  [53,] 0.19926008 0.80073992
#>  [54,] 0.39287586 0.60712414
#>  [55,] 0.39304081 0.60695919
#>  [56,] 0.27579998 0.72420002
#>  [57,] 0.27187374 0.72812626
#>  [58,] 0.29541008 0.70458992
#>  [59,] 0.28223680 0.71776320
#>  [60,] 0.20161946 0.79838054
#>  [61,] 0.21182344 0.78817656
#>  [62,] 0.25947534 0.74052466
#>  [63,] 0.67902184 0.32097816
#>  [64,] 0.67291824 0.32708176
#>  [65,] 0.57781184 0.42218816
#>  [66,] 0.68402288 0.31597712
#>  [67,] 0.58550855 0.41449145
#>  [68,] 0.78539784 0.21460216
#>  [69,] 0.70500762 0.29499238
#>  [70,] 0.62081957 0.37918043
#>  [71,] 0.69185830 0.30814170
#>  [72,] 0.83840979 0.16159021
#>  [73,] 0.91938266 0.08061734
#>  [74,] 0.91545642 0.08454358
#>  [75,] 0.78320177 0.21679823
#>  [76,] 0.83483908 0.16516092
#>  [77,] 0.89584632 0.10415368
#>  [78,] 1.00000000 0.00000000
#>  [79,] 0.72812626 0.27187374
#>  [80,] 0.67569696 0.32430304
#>  [81,] 0.78598049 0.21401951
#>  [82,] 0.71385401 0.28614599
#>  [83,] 0.78320177 0.21679823
#>  [84,] 0.91545642 0.08454358
#>  [85,] 0.79447136 0.20552864
#>  [86,] 0.88955152 0.11044848
#>  [87,] 0.78320177 0.21679823
#>  [88,] 0.88955152 0.11044848
#>  [89,] 0.81765217 0.18234783
#>  [90,] 0.88955152 0.11044848
#>  [91,] 0.88292412 0.11707588
#>  [92,] 1.00000000 0.00000000
#>  [93,] 0.88735545 0.11264455
#>  [94,] 0.79838054 0.20161946
#>  [95,] 0.88735545 0.11264455
#>  [96,] 0.91545642 0.08454358
#>  [97,] 0.69252834 0.30747166
#>  [98,] 0.69865065 0.30134935
#>  [99,] 0.79838054 0.20161946
#> [100,] 0.81130274 0.18869726
#> [101,] 0.72133391 0.27866609
#> [102,] 0.80230678 0.19769322
#> [103,] 0.69865818 0.30134182
#> [104,] 0.78320177 0.21679823
#> [105,] 0.78320177 0.21679823
#> [106,] 0.69865818 0.30134182
#> [107,] 0.82149753 0.17850247
#> [108,] 0.81130274 0.18869726
#> [109,] 0.69236339 0.30763661
#> [110,] 0.71740767 0.28259233
#> [111,] 0.79166639 0.20833361
#> [112,] 0.69865818 0.30134182
#> [113,] 0.73425610 0.26574390
#> [114,] 0.70458992 0.29541008
#> [115,] 0.88735545 0.11264455
#> [116,] 0.81130274 0.18869726
#> [117,] 0.68402288 0.31597712
#> [118,] 0.70222889 0.29777111
#> [119,] 0.79796119 0.20203881
#> [120,] 1.00000000 0.00000000
#> [121,] 0.88735545 0.11264455
#> [122,] 0.88735545 0.11264455
#> [123,] 0.88735545 0.11264455
#> [124,] 0.88735545 0.11264455
#> [125,] 0.89584632 0.10415368
#> [126,] 0.81130274 0.18869726
#> [127,] 0.81765217 0.18234783
#> [128,] 0.75386620 0.24613380
#> [129,] 0.91545642 0.08454358
#> [130,] 0.88971647 0.11028353
#> [131,] 1.00000000 0.00000000
#> [132,] 1.00000000 0.00000000
#> [133,] 0.72796130 0.27203870
#> [134,] 0.89584632 0.10415368
#> [135,] 0.88971647 0.11028353
#> [136,] 0.70736864 0.29263136
#> [137,] 0.80517289 0.19482711
#> [138,] 0.78539784 0.21460216
#> [139,] 0.73032986 0.26967014
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 2.3936193 0.0000000 5.9175521 3.8009045 0.0000000 0.0000000 2.5591484 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 2.8180785 0.0000000 0.0000000 0.0000000 0.0000000 2.5609262 2.5302542 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 0.5418216 5.3390990 0.0000000 0.6937065 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 2.0898042 0.0000000 0.0000000 0.0000000 2.7843476 5.9293652 2.3435367 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 0.0000000 2.2052901 0.0000000 1.5339792 2.4317317 1.4676358 4.6428383 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 8.6339933 1.6398788 6.0890776 4.3332526 0.0000000 5.9198906 0.0000000 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 6.2700294 3.4635715 0.0000000 0.0000000 1.6749308 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 1.3972850 0.0000000 5.9944514 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 77 62 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V47       V54       V49        V9       V36       V51       V11       V28 
#> 8.6339933 6.2700294 6.0890776 5.9944514 5.9293652 5.9198906 5.9175521 5.3390990 
#>       V45        V5       V12       V55       V17       V35       V21       V15 
#> 4.6428383 4.3332526 3.8009045 3.4635715 2.8180785 2.7843476 2.5609262 2.5591484 
#>       V22       V43        V1       V37       V40       V31       V58       V48 
#> 2.5302542 2.4317317 2.3936193 2.3435367 2.2052901 2.0898042 1.6749308 1.6398788 
#>       V42       V44        V7        V3       V27       V10       V13       V14 
#> 1.5339792 1.4676358 1.3972850 0.6937065 0.5418216 0.0000000 0.0000000 0.0000000 
#>       V16       V18       V19        V2       V20       V23       V24       V25 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V26       V29       V30       V32       V33       V34       V38       V39 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>        V4       V41       V46       V50       V52       V53       V56       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V59        V6       V60        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841