Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 59 M (0.57553957 0.42446043)  
#>   2) V12>=0.19575 79 14 M (0.82278481 0.17721519)  
#>     4) V16< 0.7258 71  8 M (0.88732394 0.11267606)  
#>       8) V42< 0.4932 62  3 M (0.95161290 0.04838710) *
#>       9) V42>=0.4932 9  4 R (0.44444444 0.55555556) *
#>     5) V16>=0.7258 8  2 R (0.25000000 0.75000000) *
#>   3) V12< 0.19575 60 15 R (0.25000000 0.75000000)  
#>     6) V4>=0.0576 12  1 M (0.91666667 0.08333333) *
#>     7) V4< 0.0576 48  4 R (0.08333333 0.91666667) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V11>=0.19645 79 13 M (0.83544304 0.16455696)  
#>     4) V33>=0.1991 61  3 M (0.95081967 0.04918033) *
#>     5) V33< 0.1991 18  8 R (0.44444444 0.55555556) *
#>   3) V11< 0.19645 60 11 R (0.18333333 0.81666667)  
#>     6) V59>=0.0109 13  4 M (0.69230769 0.30769231) *
#>     7) V59< 0.0109 47  2 R (0.04255319 0.95744681) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V48>=0.076 64 12 M (0.81250000 0.18750000)  
#>      4) V15< 0.6077 57  5 M (0.91228070 0.08771930) *
#>      5) V15>=0.6077 7  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.076 75 22 R (0.29333333 0.70666667)  
#>      6) V31< 0.385 28 10 M (0.64285714 0.35714286)  
#>       12) V19>=0.82795 14  0 M (1.00000000 0.00000000) *
#>       13) V19< 0.82795 14  4 R (0.28571429 0.71428571) *
#>      7) V31>=0.385 47  4 R (0.08510638 0.91489362)  
#>       14) V1>=0.0341 7  3 M (0.57142857 0.42857143) *
#>       15) V1< 0.0341 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V27>=0.8808 45  7 M (0.84444444 0.15555556)  
#>      4) V11>=0.198 36  0 M (1.00000000 0.00000000) *
#>      5) V11< 0.198 9  2 R (0.22222222 0.77777778) *
#>    3) V27< 0.8808 94 28 R (0.29787234 0.70212766)  
#>      6) V54>=0.02075 9  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.02075 85 19 R (0.22352941 0.77647059)  
#>       14) V20>=0.59095 25 12 M (0.52000000 0.48000000)  
#>         28) V2< 0.03165 11  1 M (0.90909091 0.09090909) *
#>         29) V2>=0.03165 14  3 R (0.21428571 0.78571429) *
#>       15) V20< 0.59095 60  6 R (0.10000000 0.90000000)  
#>         30) V33>=0.7017 7  3 M (0.57142857 0.42857143) *
#>         31) V33< 0.7017 53  2 R (0.03773585 0.96226415) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V31< 0.4513 60 16 M (0.73333333 0.26666667)  
#>      4) V30>=0.24695 46  6 M (0.86956522 0.13043478)  
#>        8) V20>=0.4667 38  0 M (1.00000000 0.00000000) *
#>        9) V20< 0.4667 8  2 R (0.25000000 0.75000000) *
#>      5) V30< 0.24695 14  4 R (0.28571429 0.71428571) *
#>    3) V31>=0.4513 79 17 R (0.21518987 0.78481013)  
#>      6) V28>=0.7982 21  7 M (0.66666667 0.33333333)  
#>       12) V10>=0.1223 13  0 M (1.00000000 0.00000000) *
#>       13) V10< 0.1223 8  1 R (0.12500000 0.87500000) *
#>      7) V28< 0.7982 58  3 R (0.05172414 0.94827586) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V51>=0.0127 85 23 M (0.72941176 0.27058824)  
#>      4) V11>=0.16725 60  7 M (0.88333333 0.11666667) *
#>      5) V11< 0.16725 25  9 R (0.36000000 0.64000000)  
#>       10) V7< 0.0398 8  0 M (1.00000000 0.00000000) *
#>       11) V7>=0.0398 17  1 R (0.05882353 0.94117647) *
#>    3) V51< 0.0127 54 10 R (0.18518519 0.81481481)  
#>      6) V42>=0.29755 15  6 M (0.60000000 0.40000000) *
#>      7) V42< 0.29755 39  1 R (0.02564103 0.97435897) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V51>=0.01365 81 20 M (0.75308642 0.24691358)  
#>      4) V48>=0.0482 71 11 M (0.84507042 0.15492958)  
#>        8) V40< 0.60865 62  4 M (0.93548387 0.06451613) *
#>        9) V40>=0.60865 9  2 R (0.22222222 0.77777778) *
#>      5) V48< 0.0482 10  1 R (0.10000000 0.90000000) *
#>    3) V51< 0.01365 58 18 R (0.31034483 0.68965517)  
#>      6) V31< 0.4008 21  5 M (0.76190476 0.23809524)  
#>       12) V60>=0.0045 14  0 M (1.00000000 0.00000000) *
#>       13) V60< 0.0045 7  2 R (0.28571429 0.71428571) *
#>      7) V31>=0.4008 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V12>=0.095 115 37 M (0.67826087 0.32173913)  
#>      4) V48>=0.077 52  4 M (0.92307692 0.07692308)  
#>        8) V25>=0.38655 45  0 M (1.00000000 0.00000000) *
#>        9) V25< 0.38655 7  3 R (0.42857143 0.57142857) *
#>      5) V48< 0.077 63 30 R (0.47619048 0.52380952)  
#>       10) V15>=0.50905 22  3 M (0.86363636 0.13636364) *
#>       11) V15< 0.50905 41 11 R (0.26829268 0.73170732)  
#>         22) V38>=0.2959 18  7 M (0.61111111 0.38888889) *
#>         23) V38< 0.2959 23  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.095 24  2 R (0.08333333 0.91666667) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V46>=0.1709 48  8 M (0.83333333 0.16666667)  
#>      4) V44>=0.1244 41  2 M (0.95121951 0.04878049) *
#>      5) V44< 0.1244 7  1 R (0.14285714 0.85714286) *
#>    3) V46< 0.1709 91 28 R (0.30769231 0.69230769)  
#>      6) V36< 0.42285 45 18 M (0.60000000 0.40000000)  
#>       12) V46>=0.06965 26  2 M (0.92307692 0.07692308) *
#>       13) V46< 0.06965 19  3 R (0.15789474 0.84210526) *
#>      7) V36>=0.42285 46  1 R (0.02173913 0.97826087) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V36< 0.53235 89 33 M (0.62921348 0.37078652)  
#>      4) V44>=0.1713 35  1 M (0.97142857 0.02857143) *
#>      5) V44< 0.1713 54 22 R (0.40740741 0.59259259)  
#>       10) V56>=0.0105 18  2 M (0.88888889 0.11111111) *
#>       11) V56< 0.0105 36  6 R (0.16666667 0.83333333)  
#>         22) V41< 0.1109 7  2 M (0.71428571 0.28571429) *
#>         23) V41>=0.1109 29  1 R (0.03448276 0.96551724) *
#>    3) V36>=0.53235 50 10 R (0.20000000 0.80000000)  
#>      6) V45>=0.28135 12  3 M (0.75000000 0.25000000) *
#>      7) V45< 0.28135 38  1 R (0.02631579 0.97368421) *
#> 
#> 
#> $weights
#>  [1] 0.7346456 0.6174010 0.6887296 0.7119043 0.7993146 0.6452194 0.8307485
#>  [8] 0.6336751 0.8630080 0.7262258
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.9524678 4.2984041
#>   [2,] 1.3599009 5.8909710
#>   [3,] 2.8938612 4.3570107
#>   [4,] 2.2779634 4.9729085
#>   [5,] 2.1140841 5.1367878
#>   [6,] 2.7948204 4.4560514
#>   [7,] 1.5517376 5.6991343
#>   [8,] 1.3455794 5.9052925
#>   [9,] 2.2688785 4.9819934
#>  [10,] 1.9972660 5.2536059
#>  [11,] 3.4480331 3.8028388
#>  [12,] 2.7091703 4.5417016
#>  [13,] 2.6744513 4.5764206
#>  [14,] 1.9513500 5.2995219
#>  [15,] 2.9898677 4.2610042
#>  [16,] 1.4006339 5.8502380
#>  [17,] 2.2313824 5.0194895
#>  [18,] 2.2916199 4.9592520
#>  [19,] 1.3798650 5.8710069
#>  [20,] 2.3419674 4.9089045
#>  [21,] 2.0570504 5.1938215
#>  [22,] 2.1707598 5.0801121
#>  [23,] 0.7119043 6.5389676
#>  [24,] 0.0000000 7.2508719
#>  [25,] 0.8307485 6.4201234
#>  [26,] 1.4167156 5.8341563
#>  [27,] 0.6336751 6.6171968
#>  [28,] 1.4966831 5.7541888
#>  [29,] 2.8953644 4.3555075
#>  [30,] 2.8280145 4.4228574
#>  [31,] 1.9513500 5.2995219
#>  [32,] 2.6938499 4.5570220
#>  [33,] 1.2626203 5.9882516
#>  [34,] 3.6235428 3.6273291
#>  [35,] 0.0000000 7.2508719
#>  [36,] 0.7993146 6.4515573
#>  [37,] 0.8630080 6.3878639
#>  [38,] 1.4966831 5.7541888
#>  [39,] 0.0000000 7.2508719
#>  [40,] 0.7993146 6.4515573
#>  [41,] 0.6336751 6.6171968
#>  [42,] 0.0000000 7.2508719
#>  [43,] 1.9676241 5.2832478
#>  [44,] 2.3011381 4.9497338
#>  [45,] 0.0000000 7.2508719
#>  [46,] 1.5082274 5.7426445
#>  [47,] 0.8630080 6.3878639
#>  [48,] 0.8630080 6.3878639
#>  [49,] 1.5255404 5.7253315
#>  [50,] 0.7262258 6.5246461
#>  [51,] 1.4381301 5.8127418
#>  [52,] 1.4644236 5.7864483
#>  [53,] 0.7119043 6.5389676
#>  [54,] 2.1676354 5.0832365
#>  [55,] 1.9629804 5.2878915
#>  [56,] 1.2510761 5.9997958
#>  [57,] 1.9398057 5.3110662
#>  [58,] 2.2106135 5.0402584
#>  [59,] 2.1646975 5.0861744
#>  [60,] 1.2510761 5.9997958
#>  [61,] 1.5339603 5.7169116
#>  [62,] 2.0782724 5.1725995
#>  [63,] 0.6887296 6.5621423
#>  [64,] 4.3682078 2.8826641
#>  [65,] 5.1222520 2.1286199
#>  [66,] 5.1059779 2.1448940
#>  [67,] 4.8539036 2.3969683
#>  [68,] 4.2429699 3.0079020
#>  [69,] 5.7426445 1.5082274
#>  [70,] 5.6759596 1.5749123
#>  [71,] 5.8502380 1.4006339
#>  [72,] 7.2508719 0.0000000
#>  [73,] 5.3545764 1.8962954
#>  [74,] 5.9284672 1.3224047
#>  [75,] 4.2086842 3.0421877
#>  [76,] 4.8308897 2.4199822
#>  [77,] 5.8127418 1.4381301
#>  [78,] 5.6759596 1.5749123
#>  [79,] 5.9072452 1.3436267
#>  [80,] 7.2508719 0.0000000
#>  [81,] 5.8909710 1.3599009
#>  [82,] 5.9215666 1.3293052
#>  [83,] 7.2508719 0.0000000
#>  [84,] 7.2508719 0.0000000
#>  [85,] 5.8027225 1.4481494
#>  [86,] 5.7704629 1.4804089
#>  [87,] 5.9447413 1.3061306
#>  [88,] 6.5162263 0.7346456
#>  [89,] 5.0050073 2.2458646
#>  [90,] 5.1706469 2.0802250
#>  [91,] 6.5621423 0.6887296
#>  [92,] 5.8909710 1.3599009
#>  [93,] 5.1059779 2.1448940
#>  [94,] 6.3878639 0.8630080
#>  [95,] 5.5885493 1.6623226
#>  [96,] 5.8341563 1.4167156
#>  [97,] 5.2536059 1.9972660
#>  [98,] 6.5389676 0.7119043
#>  [99,] 6.5162263 0.7346456
#> [100,] 5.8178822 1.4329897
#> [101,] 5.0995107 2.1513612
#> [102,] 5.6532183 1.5976536
#> [103,] 5.1412289 2.1096430
#> [104,] 5.3545764 1.8962954
#> [105,] 6.4201234 0.8307485
#> [106,] 6.4201234 0.8307485
#> [107,] 6.4201234 0.8307485
#> [108,] 5.6854778 1.5653941
#> [109,] 5.7396530 1.5112189
#> [110,] 5.7749040 1.4759679
#> [111,] 4.8861632 2.3647087
#> [112,] 5.1869210 2.0639509
#> [113,] 5.9169229 1.3339490
#> [114,] 5.7313938 1.5194781
#> [115,] 5.7313938 1.5194781
#> [116,] 5.7426445 1.5082274
#> [117,] 5.1906971 2.0601748
#> [118,] 5.0861744 2.1646975
#> [119,] 5.7313938 1.5194781
#> [120,] 6.5162263 0.7346456
#> [121,] 7.2508719 0.0000000
#> [122,] 6.5389676 0.7119043
#> [123,] 5.8127418 1.4381301
#> [124,] 5.8127418 1.4381301
#> [125,] 5.6616381 1.5892338
#> [126,] 5.6616381 1.5892338
#> [127,] 6.5246461 0.7262258
#> [128,] 5.6616381 1.5892338
#> [129,] 5.7749040 1.4759679
#> [130,] 7.2508719 0.0000000
#> [131,] 7.2508719 0.0000000
#> [132,] 7.2508719 0.0000000
#> [133,] 7.2508719 0.0000000
#> [134,] 5.9284672 1.3224047
#> [135,] 5.8027225 1.4481494
#> [136,] 5.6616381 1.5892338
#> [137,] 6.5246461 0.7262258
#> [138,] 6.5246461 0.7262258
#> [139,] 5.7749040 1.4759679
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.40718798 0.59281202
#>   [2,] 0.18754998 0.81245002
#>   [3,] 0.39910527 0.60089473
#>   [4,] 0.31416407 0.68583593
#>   [5,] 0.29156274 0.70843726
#>   [6,] 0.38544612 0.61455388
#>   [7,] 0.21400704 0.78599296
#>   [8,] 0.18557484 0.81442516
#>   [9,] 0.31291113 0.68708887
#>  [10,] 0.27545184 0.72454816
#>  [11,] 0.47553358 0.52446642
#>  [12,] 0.37363373 0.62636627
#>  [13,] 0.36884548 0.63115452
#>  [14,] 0.26911936 0.73088064
#>  [15,] 0.41234596 0.58765404
#>  [16,] 0.19316766 0.80683234
#>  [17,] 0.30773987 0.69226013
#>  [18,] 0.31604749 0.68395251
#>  [19,] 0.19030333 0.80969667
#>  [20,] 0.32299114 0.67700886
#>  [21,] 0.28369697 0.71630303
#>  [22,] 0.29937914 0.70062086
#>  [23,] 0.09818189 0.90181811
#>  [24,] 0.00000000 1.00000000
#>  [25,] 0.11457222 0.88542778
#>  [26,] 0.19538555 0.80461445
#>  [27,] 0.08739295 0.91260705
#>  [28,] 0.20641422 0.79358578
#>  [29,] 0.39931258 0.60068742
#>  [30,] 0.39002405 0.60997595
#>  [31,] 0.26911936 0.73088064
#>  [32,] 0.37152083 0.62847917
#>  [33,] 0.17413359 0.82586641
#>  [34,] 0.49973891 0.50026109
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.11023704 0.88976296
#>  [37,] 0.11902127 0.88097873
#>  [38,] 0.20641422 0.79358578
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.11023704 0.88976296
#>  [41,] 0.08739295 0.91260705
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.27136380 0.72863620
#>  [44,] 0.31736019 0.68263981
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.20800635 0.79199365
#>  [47,] 0.11902127 0.88097873
#>  [48,] 0.11902127 0.88097873
#>  [49,] 0.21039407 0.78960593
#>  [50,] 0.10015703 0.89984297
#>  [51,] 0.19833892 0.80166108
#>  [52,] 0.20196517 0.79803483
#>  [53,] 0.09818189 0.90181811
#>  [54,] 0.29894824 0.70105176
#>  [55,] 0.27072335 0.72927665
#>  [56,] 0.17254146 0.82745854
#>  [57,] 0.26752723 0.73247277
#>  [58,] 0.30487554 0.69512446
#>  [59,] 0.29854306 0.70145694
#>  [60,] 0.17254146 0.82745854
#>  [61,] 0.21155529 0.78844471
#>  [62,] 0.28662379 0.71337621
#>  [63,] 0.09498577 0.90501423
#>  [64,] 0.60243897 0.39756103
#>  [65,] 0.70643256 0.29356744
#>  [66,] 0.70418812 0.29581188
#>  [67,] 0.66942344 0.33057656
#>  [68,] 0.58516685 0.41483315
#>  [69,] 0.79199365 0.20800635
#>  [70,] 0.78279684 0.21720316
#>  [71,] 0.80683234 0.19316766
#>  [72,] 1.00000000 0.00000000
#>  [73,] 0.73847346 0.26152654
#>  [74,] 0.81762128 0.18237872
#>  [75,] 0.58043837 0.41956163
#>  [76,] 0.66624948 0.33375052
#>  [77,] 0.80166108 0.19833892
#>  [78,] 0.78279684 0.21720316
#>  [79,] 0.81469446 0.18530554
#>  [80,] 1.00000000 0.00000000
#>  [81,] 0.81245002 0.18754998
#>  [82,] 0.81666960 0.18333040
#>  [83,] 1.00000000 0.00000000
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.80027927 0.19972073
#>  [86,] 0.79583022 0.20416978
#>  [87,] 0.81986572 0.18013428
#>  [88,] 0.89868175 0.10131825
#>  [89,] 0.69026283 0.30973717
#>  [90,] 0.71310691 0.28689309
#>  [91,] 0.90501423 0.09498577
#>  [92,] 0.81245002 0.18754998
#>  [93,] 0.70418812 0.29581188
#>  [94,] 0.88097873 0.11902127
#>  [95,] 0.77074169 0.22925831
#>  [96,] 0.80461445 0.19538555
#>  [97,] 0.72454816 0.27545184
#>  [98,] 0.90181811 0.09818189
#>  [99,] 0.89868175 0.10131825
#> [100,] 0.80237001 0.19762999
#> [101,] 0.70329620 0.29670380
#> [102,] 0.77966048 0.22033952
#> [103,] 0.70904975 0.29095025
#> [104,] 0.73847346 0.26152654
#> [105,] 0.88542778 0.11457222
#> [106,] 0.88542778 0.11457222
#> [107,] 0.88542778 0.11457222
#> [108,] 0.78410953 0.21589047
#> [109,] 0.79158107 0.20841893
#> [110,] 0.79644271 0.20355729
#> [111,] 0.67387250 0.32612750
#> [112,] 0.71535135 0.28464865
#> [113,] 0.81602915 0.18397085
#> [114,] 0.79044201 0.20955799
#> [115,] 0.79044201 0.20955799
#> [116,] 0.79199365 0.20800635
#> [117,] 0.71587213 0.28412787
#> [118,] 0.70145694 0.29854306
#> [119,] 0.79044201 0.20955799
#> [120,] 0.89868175 0.10131825
#> [121,] 1.00000000 0.00000000
#> [122,] 0.90181811 0.09818189
#> [123,] 0.80166108 0.19833892
#> [124,] 0.80166108 0.19833892
#> [125,] 0.78082170 0.21917830
#> [126,] 0.78082170 0.21917830
#> [127,] 0.89984297 0.10015703
#> [128,] 0.78082170 0.21917830
#> [129,] 0.79644271 0.20355729
#> [130,] 1.00000000 0.00000000
#> [131,] 1.00000000 0.00000000
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 0.81762128 0.18237872
#> [135,] 0.80027927 0.19972073
#> [136,] 0.78082170 0.21917830
#> [137,] 0.89984297 0.10015703
#> [138,] 0.89984297 0.10015703
#> [139,] 0.79644271 0.20355729
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.7446895 1.6845747 8.4331972 7.0437487 0.0000000 0.0000000 3.7736363 1.1925099 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 1.3672033 1.1767137 2.8833604 0.0000000 0.0000000 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.6966909 0.0000000 3.5973647 2.5902960 0.0000000 0.0000000 1.6255439 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 9.2565542 0.0000000 1.9199074 0.0000000 0.0000000 6.0293700 0.0000000 1.3282073 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 2.7222563 1.8462288 1.0518981 2.1071460 0.0000000 4.6011211 1.9280001 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 7.2480873 0.0000000 7.8137833 0.0000000 0.0000000 0.0000000 6.5667970 0.0000000 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 1.9415456 0.0000000 2.5266028 0.0000000 0.0000000 1.4753607 0.0000000 
#>       V60        V7        V8        V9 
#> 1.0994175 1.7281873 0.0000000 0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 76 63 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V31       V11       V48       V46       V12       V51       V36       V44 
#> 9.2565542 8.4331972 7.8137833 7.2480873 7.0437487 6.5667970 6.0293700 4.6011211 
#>       V15       V27       V20        V4       V28       V56       V42       V54 
#> 3.7736363 3.5973647 2.8833604 2.7222563 2.5902960 2.5266028 2.1071460 1.9415456 
#>       V45       V33       V40        V7       V10       V30       V59       V19 
#> 1.9280001 1.9199074 1.8462288 1.7281873 1.6845747 1.6255439 1.4753607 1.3672033 
#>       V38       V16        V2       V60       V41        V1       V25       V13 
#> 1.3282073 1.1925099 1.1767137 1.0994175 1.0518981 0.7446895 0.6966909 0.0000000 
#>       V14       V17       V18       V21       V22       V23       V24       V26 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V29        V3       V32       V34       V35       V37       V39       V43 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V47       V49        V5       V50       V52       V53       V55       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V58        V6        V8        V9 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841