Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V9>=0.10785 87 23 M (0.73563218 0.26436782)  
#>      4) V36< 0.4098 60  6 M (0.90000000 0.10000000) *
#>      5) V36>=0.4098 27 10 R (0.37037037 0.62962963)  
#>       10) V50>=0.01195 13  3 M (0.76923077 0.23076923) *
#>       11) V50< 0.01195 14  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.10785 52  7 R (0.13461538 0.86538462)  
#>      6) V4>=0.0602 8  3 M (0.62500000 0.37500000) *
#>      7) V4< 0.0602 44  2 R (0.04545455 0.95454545) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V9>=0.1608 56 11 M (0.80357143 0.19642857)  
#>      4) V15< 0.54725 44  2 M (0.95454545 0.04545455) *
#>      5) V15>=0.54725 12  3 R (0.25000000 0.75000000) *
#>    3) V9< 0.1608 83 23 R (0.27710843 0.72289157)  
#>      6) V1>=0.05205 8  0 M (1.00000000 0.00000000) *
#>      7) V1< 0.05205 75 15 R (0.20000000 0.80000000)  
#>       14) V46>=0.1707 20  9 M (0.55000000 0.45000000)  
#>         28) V28>=0.64755 9  0 M (1.00000000 0.00000000) *
#>         29) V28< 0.64755 11  2 R (0.18181818 0.81818182) *
#>       15) V46< 0.1707 55  4 R (0.07272727 0.92727273) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V11>=0.1745 89 32 M (0.64044944 0.35955056)  
#>      4) V35< 0.3652 45  3 M (0.93333333 0.06666667) *
#>      5) V35>=0.3652 44 15 R (0.34090909 0.65909091)  
#>       10) V45>=0.27285 13  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.27285 31  2 R (0.06451613 0.93548387) *
#>    3) V11< 0.1745 50  9 R (0.18000000 0.82000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V4>=0.0576 51  4 M (0.92156863 0.07843137) *
#>    3) V4< 0.0576 88 29 R (0.32954545 0.67045455)  
#>      6) V51>=0.0137 35 12 M (0.65714286 0.34285714)  
#>       12) V49>=0.02875 27  4 M (0.85185185 0.14814815) *
#>       13) V49< 0.02875 8  0 R (0.00000000 1.00000000) *
#>      7) V51< 0.0137 53  6 R (0.11320755 0.88679245)  
#>       14) V28< 0.45545 9  3 M (0.66666667 0.33333333) *
#>       15) V28>=0.45545 44  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.5179856 0.4820144)  
#>    2) V43>=0.22215 75 17 M (0.7733333 0.2266667)  
#>      4) V11>=0.0942 67 10 M (0.8507463 0.1492537)  
#>        8) V37< 0.4665 42  0 M (1.0000000 0.0000000) *
#>        9) V37>=0.4665 25 10 M (0.6000000 0.4000000)  
#>         18) V13< 0.25735 17  2 M (0.8823529 0.1176471) *
#>         19) V13>=0.25735 8  0 R (0.0000000 1.0000000) *
#>      5) V11< 0.0942 8  1 R (0.1250000 0.8750000) *
#>    3) V43< 0.22215 64 14 R (0.2187500 0.7812500)  
#>      6) V54>=0.00905 14  3 M (0.7857143 0.2142857) *
#>      7) V54< 0.00905 50  3 R (0.0600000 0.9400000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V28>=0.9063 26  1 M (0.96153846 0.03846154) *
#>    3) V28< 0.9063 113 40 R (0.35398230 0.64601770)  
#>      6) V59>=0.00775 53 18 M (0.66037736 0.33962264)  
#>       12) V40< 0.58125 40  5 M (0.87500000 0.12500000)  
#>         24) V16>=0.28545 33  1 M (0.96969697 0.03030303) *
#>         25) V16< 0.28545 7  3 R (0.42857143 0.57142857) *
#>       13) V40>=0.58125 13  0 R (0.00000000 1.00000000) *
#>      7) V59< 0.00775 60  5 R (0.08333333 0.91666667) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V21>=0.5504 73 24 M (0.67123288 0.32876712)  
#>      4) V51>=0.0108 40  5 M (0.87500000 0.12500000) *
#>      5) V51< 0.0108 33 14 R (0.42424242 0.57575758)  
#>       10) V32< 0.31345 15  2 M (0.86666667 0.13333333) *
#>       11) V32>=0.31345 18  1 R (0.05555556 0.94444444) *
#>    3) V21< 0.5504 66 16 R (0.24242424 0.75757576)  
#>      6) V46>=0.2079 14  3 M (0.78571429 0.21428571) *
#>      7) V46< 0.2079 52  5 R (0.09615385 0.90384615)  
#>       14) V37< 0.20515 7  2 M (0.71428571 0.28571429) *
#>       15) V37>=0.20515 45  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V12>=0.1519 85 26 M (0.69411765 0.30588235)  
#>      4) V31< 0.51155 51  5 M (0.90196078 0.09803922)  
#>        8) V30>=0.302 44  1 M (0.97727273 0.02272727) *
#>        9) V30< 0.302 7  3 R (0.42857143 0.57142857) *
#>      5) V31>=0.51155 34 13 R (0.38235294 0.61764706)  
#>       10) V27>=0.71055 13  2 M (0.84615385 0.15384615) *
#>       11) V27< 0.71055 21  2 R (0.09523810 0.90476190) *
#>    3) V12< 0.1519 54  7 R (0.12962963 0.87037037)  
#>      6) V53< 0.00415 7  1 M (0.85714286 0.14285714) *
#>      7) V53>=0.00415 47  1 R (0.02127660 0.97872340) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V10>=0.1955 47  8 M (0.82978723 0.17021277)  
#>      4) V6< 0.17055 40  3 M (0.92500000 0.07500000) *
#>      5) V6>=0.17055 7  2 R (0.28571429 0.71428571) *
#>    3) V10< 0.1955 92 20 R (0.21739130 0.78260870)  
#>      6) V28< 0.4569 20  7 M (0.65000000 0.35000000)  
#>       12) V20>=0.87015 11  0 M (1.00000000 0.00000000) *
#>       13) V20< 0.87015 9  2 R (0.22222222 0.77777778) *
#>      7) V28>=0.4569 72  7 R (0.09722222 0.90277778)  
#>       14) V44>=0.2918 7  3 M (0.57142857 0.42857143) *
#>       15) V44< 0.2918 65  3 R (0.04615385 0.95384615) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V31< 0.4671 68 15 M (0.77941176 0.22058824)  
#>      4) V32>=0.16515 46  1 M (0.97826087 0.02173913) *
#>      5) V32< 0.16515 22  8 R (0.36363636 0.63636364)  
#>       10) V43>=0.25525 8  0 M (1.00000000 0.00000000) *
#>       11) V43< 0.25525 14  0 R (0.00000000 1.00000000) *
#>    3) V31>=0.4671 71 21 R (0.29577465 0.70422535)  
#>      6) V52>=0.02145 10  0 M (1.00000000 0.00000000) *
#>      7) V52< 0.02145 61 11 R (0.18032787 0.81967213)  
#>       14) V35< 0.1498 7  1 M (0.85714286 0.14285714) *
#>       15) V35>=0.1498 54  5 R (0.09259259 0.90740741) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.8165148 0.7545685 0.7910093 0.7018218 0.7107018 0.7314753
#>  [8] 0.5597176 0.6339936 0.8137003
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.3592781 4.9376641
#>   [2,] 2.1489080 5.1480342
#>   [3,] 3.0304366 4.2665056
#>   [4,] 0.7314753 6.5654669
#>   [5,] 1.8955331 5.4014091
#>   [6,] 1.4476939 5.8492483
#>   [7,] 1.5017111 5.7952311
#>   [8,] 1.2704195 6.0265227
#>   [9,] 2.3909632 4.9059790
#>  [10,] 2.2339475 5.0629947
#>  [11,] 2.0822022 5.2147400
#>  [12,] 1.2704195 6.0265227
#>  [13,] 0.7910093 6.5059329
#>  [14,] 1.4505084 5.8464338
#>  [15,] 2.8656413 4.4313009
#>  [16,] 2.0822022 5.2147400
#>  [17,] 1.3142862 5.9826560
#>  [18,] 2.2243063 5.0726358
#>  [19,] 0.7018218 6.5951204
#>  [20,] 1.2615394 6.0354027
#>  [21,] 0.7018218 6.5951204
#>  [22,] 2.0869342 5.2100080
#>  [23,] 1.5224845 5.7744577
#>  [24,] 2.3881487 4.9087935
#>  [25,] 2.3065314 4.9904108
#>  [26,] 0.7910093 6.5059329
#>  [27,] 2.0752398 5.2217024
#>  [28,] 1.3507269 5.9462153
#>  [29,] 2.0672907 5.2296515
#>  [30,] 0.5597176 6.7372246
#>  [31,] 2.1048932 5.1920490
#>  [32,] 1.4125236 5.8844185
#>  [33,] 2.0752398 5.2217024
#>  [34,] 2.3078413 4.9891009
#>  [35,] 0.7018218 6.5951204
#>  [36,] 0.7018218 6.5951204
#>  [37,] 0.7107018 6.5862404
#>  [38,] 0.5597176 6.7372246
#>  [39,] 0.7910093 6.5059329
#>  [40,] 0.0000000 7.2969422
#>  [41,] 0.0000000 7.2969422
#>  [42,] 0.0000000 7.2969422
#>  [43,] 2.8838055 4.4131367
#>  [44,] 1.2911929 6.0057493
#>  [45,] 2.3286148 4.9683274
#>  [46,] 0.7314753 6.5654669
#>  [47,] 2.3286148 4.9683274
#>  [48,] 1.5149144 5.7820278
#>  [49,] 1.5971395 5.6998027
#>  [50,] 2.3361849 4.9607573
#>  [51,] 2.9034766 4.3934656
#>  [52,] 2.3290169 4.9679253
#>  [53,] 2.7317189 4.5652233
#>  [54,] 2.8359879 4.4609543
#>  [55,] 2.8507298 4.4462124
#>  [56,] 2.3154114 4.9815308
#>  [57,] 1.4941410 5.8028012
#>  [58,] 1.5149144 5.7820278
#>  [59,] 2.2167362 5.0802060
#>  [60,] 2.0746321 5.2223101
#>  [61,] 1.4928311 5.8041111
#>  [62,] 5.0918653 2.2050769
#>  [63,] 6.6629486 0.6339936
#>  [64,] 5.8464338 1.4505084
#>  [65,] 5.9207098 1.3762324
#>  [66,] 5.1661413 2.1308009
#>  [67,] 5.0240371 2.2729051
#>  [68,] 3.8006724 3.4962698
#>  [69,] 5.8108984 1.4860438
#>  [70,] 4.5321477 2.7647945
#>  [71,] 5.9826560 1.3142862
#>  [72,] 5.7258589 1.5710833
#>  [73,] 5.2100080 2.0869342
#>  [74,] 5.8464338 1.4505084
#>  [75,] 5.7258589 1.5710833
#>  [76,] 5.7697256 1.5272166
#>  [77,] 5.8464338 1.4505084
#>  [78,] 7.2969422 0.0000000
#>  [79,] 6.5951204 0.7018218
#>  [80,] 7.2969422 0.0000000
#>  [81,] 5.8464338 1.4505084
#>  [82,] 5.9207098 1.3762324
#>  [83,] 6.4804274 0.8165148
#>  [84,] 6.4804274 0.8165148
#>  [85,] 5.3247009 1.9722413
#>  [86,] 6.5862404 0.7107018
#>  [87,] 6.4804274 0.8165148
#>  [88,] 6.1032310 1.1937112
#>  [89,] 4.4958564 2.8010858
#>  [90,] 6.4804274 0.8165148
#>  [91,] 6.4832419 0.8137003
#>  [92,] 6.0265227 1.2704195
#>  [93,] 6.7372246 0.5597176
#>  [94,] 5.9235242 1.3734180
#>  [95,] 6.5862404 0.7107018
#>  [96,] 5.2504249 2.0465173
#>  [97,] 6.5862404 0.7107018
#>  [98,] 5.1298500 2.1670922
#>  [99,] 5.1298500 2.1670922
#> [100,] 5.6922326 1.6047096
#> [101,] 5.0707182 2.2262240
#> [102,] 6.4832419 0.8137003
#> [103,] 6.4832419 0.8137003
#> [104,] 5.7725400 1.5244022
#> [105,] 4.5404789 2.7564633
#> [106,] 3.8242603 3.4726819
#> [107,] 5.8316718 1.4652704
#> [108,] 5.7589345 1.5380077
#> [109,] 6.5059329 0.7910093
#> [110,] 6.5135030 0.7834391
#> [111,] 5.0117920 2.2851502
#> [112,] 5.1001966 2.1967456
#> [113,] 5.7286733 1.5682688
#> [114,] 5.8844185 1.4125236
#> [115,] 6.5059329 0.7910093
#> [116,] 5.7744577 1.5224845
#> [117,] 6.0265227 1.2704195
#> [118,] 4.9815308 2.3154114
#> [119,] 4.9815308 2.3154114
#> [120,] 6.6629486 0.6339936
#> [121,] 7.2969422 0.0000000
#> [122,] 7.2969422 0.0000000
#> [123,] 6.6629486 0.6339936
#> [124,] 5.2355135 2.0614287
#> [125,] 6.4832419 0.8137003
#> [126,] 5.6667271 1.6302151
#> [127,] 5.1474264 2.1495157
#> [128,] 4.3639873 2.9329549
#> [129,] 5.1474264 2.1495157
#> [130,] 5.9611268 1.3358154
#> [131,] 7.2969422 0.0000000
#> [132,] 6.5951204 0.7018218
#> [133,] 7.2969422 0.0000000
#> [134,] 6.5951204 0.7018218
#> [135,] 5.8041111 1.4928311
#> [136,] 6.5951204 0.7018218
#> [137,] 7.2969422 0.0000000
#> [138,] 6.5951204 0.7018218
#> [139,] 5.8547651 1.4421771
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32332422 0.67667578
#>   [2,] 0.29449432 0.70550568
#>   [3,] 0.41530226 0.58469774
#>   [4,] 0.10024408 0.89975592
#>   [5,] 0.25977087 0.74022913
#>   [6,] 0.19839734 0.80160266
#>   [7,] 0.20580005 0.79419995
#>   [8,] 0.17410299 0.82589701
#>   [9,] 0.32766646 0.67233354
#>  [10,] 0.30614845 0.69385155
#>  [11,] 0.28535270 0.71464730
#>  [12,] 0.17410299 0.82589701
#>  [13,] 0.10840284 0.89159716
#>  [14,] 0.19878304 0.80121696
#>  [15,] 0.39271811 0.60728189
#>  [16,] 0.28535270 0.71464730
#>  [17,] 0.18011465 0.81988535
#>  [18,] 0.30482718 0.69517282
#>  [19,] 0.09618026 0.90381974
#>  [20,] 0.17288604 0.82711396
#>  [21,] 0.09618026 0.90381974
#>  [22,] 0.28600120 0.71399880
#>  [23,] 0.20864692 0.79135308
#>  [24,] 0.32728075 0.67271925
#>  [25,] 0.31609561 0.68390439
#>  [26,] 0.10840284 0.89159716
#>  [27,] 0.28439855 0.71560145
#>  [28,] 0.18510862 0.81489138
#>  [29,] 0.28330918 0.71669082
#>  [30,] 0.07670578 0.92329422
#>  [31,] 0.28846237 0.71153763
#>  [32,] 0.19357748 0.80642252
#>  [33,] 0.28439855 0.71560145
#>  [34,] 0.31627513 0.68372487
#>  [35,] 0.09618026 0.90381974
#>  [36,] 0.09618026 0.90381974
#>  [37,] 0.09739721 0.90260279
#>  [38,] 0.07670578 0.92329422
#>  [39,] 0.10840284 0.89159716
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.39520739 0.60479261
#>  [44,] 0.17694986 0.82305014
#>  [45,] 0.31912199 0.68087801
#>  [46,] 0.10024408 0.89975592
#>  [47,] 0.31912199 0.68087801
#>  [48,] 0.20760949 0.79239051
#>  [49,] 0.21887791 0.78112209
#>  [50,] 0.32015943 0.67984057
#>  [51,] 0.39790319 0.60209681
#>  [52,] 0.31917711 0.68082289
#>  [53,] 0.37436488 0.62563512
#>  [54,] 0.38865429 0.61134571
#>  [55,] 0.39067458 0.60932542
#>  [56,] 0.31731256 0.68268744
#>  [57,] 0.20476262 0.79523738
#>  [58,] 0.20760949 0.79239051
#>  [59,] 0.30378975 0.69621025
#>  [60,] 0.28431527 0.71568473
#>  [61,] 0.20458310 0.79541690
#>  [62,] 0.69780809 0.30219191
#>  [63,] 0.91311517 0.08688483
#>  [64,] 0.80121696 0.19878304
#>  [65,] 0.81139601 0.18860399
#>  [66,] 0.70798714 0.29201286
#>  [67,] 0.68851266 0.31148734
#>  [68,] 0.52085823 0.47914177
#>  [69,] 0.79634705 0.20365295
#>  [70,] 0.62110231 0.37889769
#>  [71,] 0.81988535 0.18011465
#>  [72,] 0.78469292 0.21530708
#>  [73,] 0.71399880 0.28600120
#>  [74,] 0.80121696 0.19878304
#>  [75,] 0.78469292 0.21530708
#>  [76,] 0.79070458 0.20929542
#>  [77,] 0.80121696 0.19878304
#>  [78,] 1.00000000 0.00000000
#>  [79,] 0.90381974 0.09618026
#>  [80,] 1.00000000 0.00000000
#>  [81,] 0.80121696 0.19878304
#>  [82,] 0.81139601 0.18860399
#>  [83,] 0.88810179 0.11189821
#>  [84,] 0.88810179 0.11189821
#>  [85,] 0.72971675 0.27028325
#>  [86,] 0.90260279 0.09739721
#>  [87,] 0.88810179 0.11189821
#>  [88,] 0.83640939 0.16359061
#>  [89,] 0.61612882 0.38387118
#>  [90,] 0.88810179 0.11189821
#>  [91,] 0.88848749 0.11151251
#>  [92,] 0.82589701 0.17410299
#>  [93,] 0.92329422 0.07670578
#>  [94,] 0.81178171 0.18821829
#>  [95,] 0.90260279 0.09739721
#>  [96,] 0.71953769 0.28046231
#>  [97,] 0.90260279 0.09739721
#>  [98,] 0.70301366 0.29698634
#>  [99,] 0.70301366 0.29698634
#> [100,] 0.78008465 0.21991535
#> [101,] 0.69491002 0.30508998
#> [102,] 0.88848749 0.11151251
#> [103,] 0.88848749 0.11151251
#> [104,] 0.79109028 0.20890972
#> [105,] 0.62224406 0.37775594
#> [106,] 0.52409080 0.47590920
#> [107,] 0.79919392 0.20080608
#> [108,] 0.78922573 0.21077427
#> [109,] 0.89159716 0.10840284
#> [110,] 0.89263460 0.10736540
#> [111,] 0.68683454 0.31316546
#> [112,] 0.69894984 0.30105016
#> [113,] 0.78507863 0.21492137
#> [114,] 0.80642252 0.19357748
#> [115,] 0.89159716 0.10840284
#> [116,] 0.79135308 0.20864692
#> [117,] 0.82589701 0.17410299
#> [118,] 0.68268744 0.31731256
#> [119,] 0.68268744 0.31731256
#> [120,] 0.91311517 0.08688483
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 0.91311517 0.08688483
#> [124,] 0.71749417 0.28250583
#> [125,] 0.88848749 0.11151251
#> [126,] 0.77658928 0.22341072
#> [127,] 0.70542240 0.29457760
#> [128,] 0.59805699 0.40194301
#> [129,] 0.70542240 0.29457760
#> [130,] 0.81693490 0.18306510
#> [131,] 1.00000000 0.00000000
#> [132,] 0.90381974 0.09618026
#> [133,] 1.00000000 0.00000000
#> [134,] 0.90381974 0.09618026
#> [135,] 0.79541690 0.20458310
#> [136,] 0.90381974 0.09618026
#> [137,] 1.00000000 0.00000000
#> [138,] 0.90381974 0.09618026
#> [139,] 0.80235870 0.19764130
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 2.0555149 4.0246099 4.2243203 3.2046439 1.6173867 0.0000000 2.0793683 0.6539415 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 0.0000000 0.0000000 1.0330101 2.5367761 0.0000000 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 1.3789185 7.5686148 0.0000000 0.0000000 0.5537643 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 5.2746261 4.6316615 0.0000000 0.0000000 4.8096178 2.2266349 2.1876964 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 5.8400234 2.9049219 0.0000000 0.0000000 6.3099778 0.6015064 3.2909941 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 3.5718777 0.0000000 0.0000000 1.9275199 0.0000000 1.7003160 4.1466726 2.5557582 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 1.2964273 2.1997823 0.0000000 0.0000000 0.0000000 0.0000000 3.6237435 0.8399211 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 9.1294520 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 78 61 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V9       V28       V43        V4       V31       V35       V32       V11 
#> 9.1294520 7.5686148 6.3099778 5.8400234 5.2746261 4.8096178 4.6316615 4.2243203 
#>       V51       V10       V59       V46       V45       V12       V40       V52 
#> 4.1466726 4.0246099 3.6237435 3.5718777 3.2909941 3.2046439 2.9049219 2.5557582 
#>       V21       V36       V54       V37       V15        V1       V49       V50 
#> 2.5367761 2.2266349 2.1997823 2.1876964 2.0793683 2.0555149 1.9275199 1.7003160 
#>       V13       V27       V53       V20        V6       V16       V44       V30 
#> 1.6173867 1.3789185 1.2964273 1.0330101 0.8399211 0.6539415 0.6015064 0.5537643 
#>       V14       V17       V18       V19        V2       V22       V23       V24 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V25       V26       V29        V3       V33       V34       V38       V39 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V41       V42       V47       V48        V5       V55       V56       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V58       V60        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986