Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V4>=0.0606 47  2 M (0.95744681 0.04255319) *
#>    3) V4< 0.0606 92 30 R (0.32608696 0.67391304)  
#>      6) V10>=0.1978 27  6 M (0.77777778 0.22222222)  
#>       12) V28>=0.61435 19  0 M (1.00000000 0.00000000) *
#>       13) V28< 0.61435 8  2 R (0.25000000 0.75000000) *
#>      7) V10< 0.1978 65  9 R (0.13846154 0.86153846)  
#>       14) V28>=0.9403 7  1 M (0.85714286 0.14285714) *
#>       15) V28< 0.9403 58  3 R (0.05172414 0.94827586) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V10>=0.19155 50  8 M (0.84000000 0.16000000)  
#>      4) V48>=0.07545 38  0 M (1.00000000 0.00000000) *
#>      5) V48< 0.07545 12  4 R (0.33333333 0.66666667) *
#>    3) V10< 0.19155 89 24 R (0.26966292 0.73033708)  
#>      6) V54>=0.01685 9  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.01685 80 15 R (0.18750000 0.81250000)  
#>       14) V28>=0.958 12  3 M (0.75000000 0.25000000) *
#>       15) V28< 0.958 68  6 R (0.08823529 0.91176471) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 54 M (0.61151079 0.38848921)  
#>   2) V52>=0.00945 83 12 M (0.85542169 0.14457831)  
#>     4) V11>=0.1117 74  6 M (0.91891892 0.08108108) *
#>     5) V11< 0.1117 9  3 R (0.33333333 0.66666667) *
#>   3) V52< 0.00945 56 14 R (0.25000000 0.75000000)  
#>     6) V6>=0.10645 15  5 M (0.66666667 0.33333333) *
#>     7) V6< 0.10645 41  4 R (0.09756098 0.90243902) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 M (0.54676259 0.45323741)  
#>   2) V36< 0.4926 83 20 M (0.75903614 0.24096386)  
#>     4) V22>=0.5777 65  7 M (0.89230769 0.10769231)  
#>       8) V43>=0.0862 57  2 M (0.96491228 0.03508772) *
#>       9) V43< 0.0862 8  3 R (0.37500000 0.62500000) *
#>     5) V22< 0.5777 18  5 R (0.27777778 0.72222222) *
#>   3) V36>=0.4926 56 13 R (0.23214286 0.76785714)  
#>     6) V45>=0.2797 16  4 M (0.75000000 0.25000000) *
#>     7) V45< 0.2797 40  1 R (0.02500000 0.97500000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V36< 0.53515 103 30 M (0.70873786 0.29126214)  
#>      4) V51>=0.01265 57  3 M (0.94736842 0.05263158) *
#>      5) V51< 0.01265 46 19 R (0.41304348 0.58695652)  
#>       10) V23>=0.7036 19  3 M (0.84210526 0.15789474) *
#>       11) V23< 0.7036 27  3 R (0.11111111 0.88888889) *
#>    3) V36>=0.53515 36  4 R (0.11111111 0.88888889) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V9>=0.13175 82 23 M (0.71951220 0.28048780)  
#>      4) V49>=0.0231 56  6 M (0.89285714 0.10714286)  
#>        8) V37< 0.5221 46  0 M (1.00000000 0.00000000) *
#>        9) V37>=0.5221 10  4 R (0.40000000 0.60000000) *
#>      5) V49< 0.0231 26  9 R (0.34615385 0.65384615)  
#>       10) V28< 0.71115 9  0 M (1.00000000 0.00000000) *
#>       11) V28>=0.71115 17  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.13175 57 12 R (0.21052632 0.78947368)  
#>      6) V58>=0.0098 11  3 M (0.72727273 0.27272727) *
#>      7) V58< 0.0098 46  4 R (0.08695652 0.91304348) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 M (0.61870504 0.38129496)  
#>    2) V11>=0.1687 89 18 M (0.79775281 0.20224719)  
#>      4) V51>=0.01195 62  3 M (0.95161290 0.04838710) *
#>      5) V51< 0.01195 27 12 R (0.44444444 0.55555556)  
#>       10) V19< 0.5844 13  1 M (0.92307692 0.07692308) *
#>       11) V19>=0.5844 14  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1687 50 15 R (0.30000000 0.70000000)  
#>      6) V32< 0.36405 20  6 M (0.70000000 0.30000000)  
#>       12) V20>=0.8255 11  0 M (1.00000000 0.00000000) *
#>       13) V20< 0.8255 9  3 R (0.33333333 0.66666667) *
#>      7) V32>=0.36405 30  1 R (0.03333333 0.96666667) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V39>=0.16095 115 36 M (0.68695652 0.31304348)  
#>      4) V52>=0.007 98 22 M (0.77551020 0.22448980)  
#>        8) V48>=0.02615 85 13 M (0.84705882 0.15294118)  
#>         16) V37< 0.85895 78  7 M (0.91025641 0.08974359)  
#>           32) V11>=0.1779 53  0 M (1.00000000 0.00000000) *
#>           33) V11< 0.1779 25  7 M (0.72000000 0.28000000)  
#>             66) V45>=0.11775 16  1 M (0.93750000 0.06250000) *
#>             67) V45< 0.11775 9  3 R (0.33333333 0.66666667) *
#>         17) V37>=0.85895 7  1 R (0.14285714 0.85714286) *
#>        9) V48< 0.02615 13  4 R (0.30769231 0.69230769) *
#>      5) V52< 0.007 17  3 R (0.17647059 0.82352941) *
#>    3) V39< 0.16095 24  3 R (0.12500000 0.87500000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V45>=0.2124 31  1 M (0.96774194 0.03225806) *
#>    3) V45< 0.2124 108 48 R (0.44444444 0.55555556)  
#>      6) V37< 0.4645 71 27 M (0.61971831 0.38028169)  
#>       12) V10>=0.1653 37  3 M (0.91891892 0.08108108) *
#>       13) V10< 0.1653 34 10 R (0.29411765 0.70588235)  
#>         26) V28< 0.58085 11  2 M (0.81818182 0.18181818) *
#>         27) V28>=0.58085 23  1 R (0.04347826 0.95652174) *
#>      7) V37>=0.4645 37  4 R (0.10810811 0.89189189)  
#>       14) V1>=0.04405 7  3 M (0.57142857 0.42857143) *
#>       15) V1< 0.04405 30  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V54>=0.01855 25  0 M (1.00000000 0.00000000) *
#>    3) V54< 0.01855 114 40 R (0.35087719 0.64912281)  
#>      6) V28>=0.9233 22  3 M (0.86363636 0.13636364) *
#>      7) V28< 0.9233 92 21 R (0.22826087 0.77173913)  
#>       14) V53>=0.017 12  3 M (0.75000000 0.25000000) *
#>       15) V53< 0.017 80 12 R (0.15000000 0.85000000)  
#>         30) V3< 0.0169 15  7 M (0.53333333 0.46666667) *
#>         31) V3>=0.0169 65  4 R (0.06153846 0.93846154)  
#>           62) V50>=0.03255 7  3 M (0.57142857 0.42857143) *
#>           63) V50< 0.03255 58  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.8090480 0.6712565 0.6940692 0.7594661 0.7073843 0.6956054 0.7530688
#>  [8] 0.6356528 0.8870923 0.5927335
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.3673969 4.8379801
#>   [2,] 2.8982360 4.3071410
#>   [3,] 2.2301295 4.9752475
#>   [4,] 1.5826978 5.6226792
#>   [5,] 1.5826978 5.6226792
#>   [6,] 2.0462688 5.1591082
#>   [7,] 0.5927335 6.6126435
#>   [8,] 3.6997451 3.5056320
#>   [9,] 0.6956054 6.5097716
#>  [10,] 2.2561859 4.9491911
#>  [11,] 1.5031171 5.7022599
#>  [12,] 1.9239918 5.2813852
#>  [13,] 2.3902095 4.8151675
#>  [14,] 2.9269638 4.2784132
#>  [15,] 0.6956054 6.5097716
#>  [16,] 1.4798259 5.7255511
#>  [17,] 1.3069094 5.8984676
#>  [18,] 3.2086752 3.9967019
#>  [19,] 0.6940692 6.5113079
#>  [20,] 2.2168143 4.9885627
#>  [21,] 2.1754313 5.0299457
#>  [22,] 0.8090480 6.3963290
#>  [23,] 2.1545223 5.0508547
#>  [24,] 1.2868027 5.9185743
#>  [25,] 1.4029898 5.8023872
#>  [26,] 1.4014535 5.8039235
#>  [27,] 1.4798259 5.7255511
#>  [28,] 2.7354769 4.4699001
#>  [29,] 2.9778167 4.2275603
#>  [30,] 1.3896746 5.8157024
#>  [31,] 1.5944767 5.6109003
#>  [32,] 1.4471379 5.7582391
#>  [33,] 2.2392920 4.9660851
#>  [34,] 0.0000000 7.2053770
#>  [35,] 0.7594661 6.4459109
#>  [36,] 1.4029898 5.8023872
#>  [37,] 2.9143164 4.2910606
#>  [38,] 0.0000000 7.2053770
#>  [39,] 0.5927335 6.6126435
#>  [40,] 0.7073843 6.4979927
#>  [41,] 1.3001179 5.9052591
#>  [42,] 1.3001179 5.9052591
#>  [43,] 2.0730381 5.1323389
#>  [44,] 0.5927335 6.6126435
#>  [45,] 0.5927335 6.6126435
#>  [46,] 1.9595955 5.2457815
#>  [47,] 0.5927335 6.6126435
#>  [48,] 1.3297220 5.8756550
#>  [49,] 0.8870923 6.3182847
#>  [50,] 0.6940692 6.5113079
#>  [51,] 0.5927335 6.6126435
#>  [52,] 0.5927335 6.6126435
#>  [53,] 1.3001179 5.9052591
#>  [54,] 1.3001179 5.9052591
#>  [55,] 1.3001179 5.9052591
#>  [56,] 0.7073843 6.4979927
#>  [57,] 1.4668504 5.7385266
#>  [58,] 1.9941871 5.2111900
#>  [59,] 1.5811615 5.6242155
#>  [60,] 3.0419426 4.1634344
#>  [61,] 2.1427434 5.0626337
#>  [62,] 2.0599781 5.1453989
#>  [63,] 0.7073843 6.4979927
#>  [64,] 0.7530688 6.4523083
#>  [65,] 0.6956054 6.5097716
#>  [66,] 4.3927243 2.8126527
#>  [67,] 6.4459109 0.7594661
#>  [68,] 6.4459109 0.7594661
#>  [69,] 5.7746544 1.4307226
#>  [70,] 5.8531774 1.3521996
#>  [71,] 5.0666070 2.1387700
#>  [72,] 4.2646194 2.9407576
#>  [73,] 4.9286101 2.2767669
#>  [74,] 5.0854463 2.1199307
#>  [75,] 5.0666070 2.1387700
#>  [76,] 5.8741187 1.3312583
#>  [77,] 5.0542946 2.1510824
#>  [78,] 5.1028738 2.1025032
#>  [79,] 6.3963290 0.8090480
#>  [80,] 4.6101287 2.5952484
#>  [81,] 6.3963290 0.8090480
#>  [82,] 7.2053770 0.0000000
#>  [83,] 6.3182847 0.8870923
#>  [84,] 6.5341205 0.6712565
#>  [85,] 4.2929573 2.9124198
#>  [86,] 5.1323389 2.0730381
#>  [87,] 5.9769906 1.2283864
#>  [88,] 5.8984676 1.3069094
#>  [89,] 5.8035955 1.4017815
#>  [90,] 6.5697242 0.6356528
#>  [91,] 6.5341205 0.6712565
#>  [92,] 6.5341205 0.6712565
#>  [93,] 4.9933392 2.2120378
#>  [94,] 5.8023872 1.4029898
#>  [95,] 5.8023872 1.4029898
#>  [96,] 4.4954779 2.7098992
#>  [97,] 5.0429212 2.1624559
#>  [98,] 7.2053770 0.0000000
#>  [99,] 6.4459109 0.7594661
#> [100,] 6.5341205 0.6712565
#> [101,] 7.2053770 0.0000000
#> [102,] 5.0894197 2.1159574
#> [103,] 6.4459109 0.7594661
#> [104,] 5.9413869 1.2639901
#> [105,] 5.1161890 2.0891881
#> [106,] 4.3299536 2.8754234
#> [107,] 5.8035955 1.4017815
#> [108,] 5.2175246 1.9878524
#> [109,] 5.3057341 1.8996429
#> [110,] 4.3953505 2.8100265
#> [111,] 5.8023872 1.4029898
#> [112,] 5.8023872 1.4029898
#> [113,] 5.8023872 1.4029898
#> [114,] 6.6126435 0.5927335
#> [115,] 4.9491911 2.2561859
#> [116,] 4.9972367 2.2081403
#> [117,] 4.3552493 2.8501277
#> [118,] 5.8023872 1.4029898
#> [119,] 5.1083181 2.0970589
#> [120,] 4.8057498 2.3996272
#> [121,] 5.0672700 2.1381070
#> [122,] 5.3057341 1.8996429
#> [123,] 5.1028738 2.1025032
#> [124,] 5.9052591 1.3001179
#> [125,] 7.2053770 0.0000000
#> [126,] 7.2053770 0.0000000
#> [127,] 7.2053770 0.0000000
#> [128,] 5.7385266 1.4668504
#> [129,] 6.4459109 0.7594661
#> [130,] 5.7503055 1.4550715
#> [131,] 6.5697242 0.6356528
#> [132,] 6.5697242 0.6356528
#> [133,] 5.8984676 1.3069094
#> [134,] 6.5697242 0.6356528
#> [135,] 6.4459109 0.7594661
#> [136,] 7.2053770 0.0000000
#> [137,] 6.4459109 0.7594661
#> [138,] 5.9769906 1.2283864
#> [139,] 5.2028622 2.0025148
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32855975 0.67144025
#>   [2,] 0.40223239 0.59776761
#>   [3,] 0.30950907 0.69049093
#>   [4,] 0.21965509 0.78034491
#>   [5,] 0.21965509 0.78034491
#>   [6,] 0.28399191 0.71600809
#>   [7,] 0.08226267 0.91773733
#>   [8,] 0.51347002 0.48652998
#>   [9,] 0.09653977 0.90346023
#>  [10,] 0.31312531 0.68687469
#>  [11,] 0.20861048 0.79138952
#>  [12,] 0.26702167 0.73297833
#>  [13,] 0.33172580 0.66827420
#>  [14,] 0.40621938 0.59378062
#>  [15,] 0.09653977 0.90346023
#>  [16,] 0.20537799 0.79462201
#>  [17,] 0.18137973 0.81862027
#>  [18,] 0.44531676 0.55468324
#>  [19,] 0.09632656 0.90367344
#>  [20,] 0.30766111 0.69233889
#>  [21,] 0.30191776 0.69808224
#>  [22,] 0.11228392 0.88771608
#>  [23,] 0.29901590 0.70098410
#>  [24,] 0.17858923 0.82141077
#>  [25,] 0.19471428 0.80528572
#>  [26,] 0.19450107 0.80549893
#>  [27,] 0.20537799 0.79462201
#>  [28,] 0.37964383 0.62035617
#>  [29,] 0.41327701 0.58672299
#>  [30,] 0.19286633 0.80713367
#>  [31,] 0.22128983 0.77871017
#>  [32,] 0.20084139 0.79915861
#>  [33,] 0.31078068 0.68921932
#>  [34,] 0.00000000 1.00000000
#>  [35,] 0.10540268 0.89459732
#>  [36,] 0.19471428 0.80528572
#>  [37,] 0.40446411 0.59553589
#>  [38,] 0.00000000 1.00000000
#>  [39,] 0.08226267 0.91773733
#>  [40,] 0.09817451 0.90182549
#>  [41,] 0.18043718 0.81956282
#>  [42,] 0.18043718 0.81956282
#>  [43,] 0.28770709 0.71229291
#>  [44,] 0.08226267 0.91773733
#>  [45,] 0.08226267 0.91773733
#>  [46,] 0.27196294 0.72803706
#>  [47,] 0.08226267 0.91773733
#>  [48,] 0.18454579 0.81545421
#>  [49,] 0.12311532 0.87688468
#>  [50,] 0.09632656 0.90367344
#>  [51,] 0.08226267 0.91773733
#>  [52,] 0.08226267 0.91773733
#>  [53,] 0.18043718 0.81956282
#>  [54,] 0.18043718 0.81956282
#>  [55,] 0.18043718 0.81956282
#>  [56,] 0.09817451 0.90182549
#>  [57,] 0.20357719 0.79642281
#>  [58,] 0.27676373 0.72323627
#>  [59,] 0.21944188 0.78055812
#>  [60,] 0.42217675 0.57782325
#>  [61,] 0.29738116 0.70261884
#>  [62,] 0.28589457 0.71410543
#>  [63,] 0.09817451 0.90182549
#>  [64,] 0.10451483 0.89548517
#>  [65,] 0.09653977 0.90346023
#>  [66,] 0.60964531 0.39035469
#>  [67,] 0.89459732 0.10540268
#>  [68,] 0.89459732 0.10540268
#>  [69,] 0.80143681 0.19856319
#>  [70,] 0.81233465 0.18766535
#>  [71,] 0.70317029 0.29682971
#>  [72,] 0.59186624 0.40813376
#>  [73,] 0.68401835 0.31598165
#>  [74,] 0.70578490 0.29421510
#>  [75,] 0.70317029 0.29682971
#>  [76,] 0.81524100 0.18475900
#>  [77,] 0.70146150 0.29853850
#>  [78,] 0.70820358 0.29179642
#>  [79,] 0.88771608 0.11228392
#>  [80,] 0.63981783 0.36018217
#>  [81,] 0.88771608 0.11228392
#>  [82,] 1.00000000 0.00000000
#>  [83,] 0.87688468 0.12311532
#>  [84,] 0.90683950 0.09316050
#>  [85,] 0.59579912 0.40420088
#>  [86,] 0.71229291 0.28770709
#>  [87,] 0.82951810 0.17048190
#>  [88,] 0.81862027 0.18137973
#>  [89,] 0.80545341 0.19454659
#>  [90,] 0.91178077 0.08821923
#>  [91,] 0.90683950 0.09316050
#>  [92,] 0.90683950 0.09316050
#>  [93,] 0.69300180 0.30699820
#>  [94,] 0.80528572 0.19471428
#>  [95,] 0.80528572 0.19471428
#>  [96,] 0.62390599 0.37609401
#>  [97,] 0.69988304 0.30011696
#>  [98,] 1.00000000 0.00000000
#>  [99,] 0.89459732 0.10540268
#> [100,] 0.90683950 0.09316050
#> [101,] 1.00000000 0.00000000
#> [102,] 0.70633634 0.29366366
#> [103,] 0.89459732 0.10540268
#> [104,] 0.82457683 0.17542317
#> [105,] 0.71005153 0.28994847
#> [106,] 0.60093366 0.39906634
#> [107,] 0.80545341 0.19454659
#> [108,] 0.72411541 0.27588459
#> [109,] 0.73635760 0.26364240
#> [110,] 0.61000979 0.38999021
#> [111,] 0.80528572 0.19471428
#> [112,] 0.80528572 0.19471428
#> [113,] 0.80528572 0.19471428
#> [114,] 0.91773733 0.08226267
#> [115,] 0.68687469 0.31312531
#> [116,] 0.69354272 0.30645728
#> [117,] 0.60444433 0.39555567
#> [118,] 0.80528572 0.19471428
#> [119,] 0.70895916 0.29104084
#> [120,] 0.66696716 0.33303284
#> [121,] 0.70326230 0.29673770
#> [122,] 0.73635760 0.26364240
#> [123,] 0.70820358 0.29179642
#> [124,] 0.81956282 0.18043718
#> [125,] 1.00000000 0.00000000
#> [126,] 1.00000000 0.00000000
#> [127,] 1.00000000 0.00000000
#> [128,] 0.79642281 0.20357719
#> [129,] 0.89459732 0.10540268
#> [130,] 0.79805755 0.20194245
#> [131,] 0.91178077 0.08821923
#> [132,] 0.91178077 0.08821923
#> [133,] 0.81862027 0.18137973
#> [134,] 0.91178077 0.08821923
#> [135,] 0.89459732 0.10540268
#> [136,] 1.00000000 0.00000000
#> [137,] 0.89459732 0.10540268
#> [138,] 0.82951810 0.17048190
#> [139,] 0.72208050 0.27791950
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "M" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.9414320 11.1287346  4.9988411  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  0.0000000  0.0000000  2.4768329  0.0000000  0.9487155  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  2.3151593  2.4139121  0.0000000  0.0000000  0.0000000  0.0000000 12.1067097 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.9207871  0.0000000  0.0000000  2.2999163  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  7.8966258  5.7890206  0.0000000  2.2824347  5.7448377  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  1.0617283  0.0000000  6.7283424  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  2.7520392  2.1139395  0.0000000  0.6922688  5.0303492  6.7636293  1.2750408 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  4.9852380  0.0000000  0.0000000  0.0000000  1.4497940  0.0000000  1.4136906 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  3.4699804 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V28        V10        V36        V52        V45        V37         V4 
#> 12.1067097 11.1287346  7.8966258  6.7636293  6.7283424  5.7890206  5.7448377 
#>        V51        V11        V54         V9        V48        V19        V23 
#>  5.0303492  4.9988411  4.9852380  3.4699804  2.7520392  2.4768329  2.4139121 
#>        V22        V32        V39        V49        V58         V6        V53 
#>  2.3151593  2.2999163  2.2824347  2.1139395  1.4497940  1.4136906  1.2750408 
#>        V43        V20         V1         V3        V50        V12        V13 
#>  1.0617283  0.9487155  0.9414320  0.9207871  0.6922688  0.0000000  0.0000000 
#>        V14        V15        V16        V17        V18         V2        V21 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V25        V26        V27        V29        V30        V31 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V33        V34        V35        V38        V40        V41        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V44        V46        V47         V5        V55        V56        V57 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841