Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V11>=0.1788 87 17 M (0.80459770 0.19540230)  
#>      4) V51>=0.0123 65  5 M (0.92307692 0.07692308) *
#>      5) V51< 0.0123 22 10 R (0.45454545 0.54545455)  
#>       10) V55< 0.0067 11  1 M (0.90909091 0.09090909) *
#>       11) V55>=0.0067 11  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1788 52  6 R (0.11538462 0.88461538) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V9>=0.1164 86 28 M (0.67441860 0.32558140)  
#>      4) V37< 0.37295 57  9 M (0.84210526 0.15789474)  
#>        8) V46>=0.06695 48  2 M (0.95833333 0.04166667) *
#>        9) V46< 0.06695 9  2 R (0.22222222 0.77777778) *
#>      5) V37>=0.37295 29 10 R (0.34482759 0.65517241)  
#>       10) V45>=0.2682 7  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.2682 22  3 R (0.13636364 0.86363636) *
#>    3) V9< 0.1164 53  9 R (0.16981132 0.83018868)  
#>      6) V47>=0.1718 7  1 M (0.85714286 0.14285714) *
#>      7) V47< 0.1718 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V21>=0.63915 72 19 M (0.73611111 0.26388889)  
#>      4) V42>=0.23335 40  2 M (0.95000000 0.05000000) *
#>      5) V42< 0.23335 32 15 R (0.46875000 0.53125000)  
#>       10) V16>=0.35325 19  5 M (0.73684211 0.26315789) *
#>       11) V16< 0.35325 13  1 R (0.07692308 0.92307692) *
#>    3) V21< 0.63915 67 20 R (0.29850746 0.70149254)  
#>      6) V46>=0.1734 26  7 M (0.73076923 0.26923077)  
#>       12) V23< 0.5539 19  1 M (0.94736842 0.05263158) *
#>       13) V23>=0.5539 7  1 R (0.14285714 0.85714286) *
#>      7) V46< 0.1734 41  1 R (0.02439024 0.97560976) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V12>=0.20075 97 25 M (0.74226804 0.25773196)  
#>      4) V5>=0.0131 90 18 M (0.80000000 0.20000000)  
#>        8) V53>=0.00375 76 10 M (0.86842105 0.13157895)  
#>         16) V34< 0.6809 69  5 M (0.92753623 0.07246377) *
#>         17) V34>=0.6809 7  2 R (0.28571429 0.71428571) *
#>        9) V53< 0.00375 14  6 R (0.42857143 0.57142857) *
#>      5) V5< 0.0131 7  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.20075 42  7 R (0.16666667 0.83333333)  
#>      6) V27< 0.5952 11  5 M (0.54545455 0.45454545) *
#>      7) V27>=0.5952 31  1 R (0.03225806 0.96774194) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.5107914 0.4892086)  
#>    2) V47>=0.08395 92 31 M (0.6630435 0.3369565)  
#>      4) V4>=0.03485 58  8 M (0.8620690 0.1379310)  
#>        8) V16< 0.7239 50  1 M (0.9800000 0.0200000) *
#>        9) V16>=0.7239 8  1 R (0.1250000 0.8750000) *
#>      5) V4< 0.03485 34 11 R (0.3235294 0.6764706)  
#>       10) V14< 0.1044 9  0 M (1.0000000 0.0000000) *
#>       11) V14>=0.1044 25  2 R (0.0800000 0.9200000) *
#>    3) V47< 0.08395 47 10 R (0.2127660 0.7872340)  
#>      6) V24< 0.42575 8  3 M (0.6250000 0.3750000) *
#>      7) V24>=0.42575 39  5 R (0.1282051 0.8717949)  
#>       14) V54>=0.01225 7  3 M (0.5714286 0.4285714) *
#>       15) V54< 0.01225 32  1 R (0.0312500 0.9687500) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V37< 0.1872 42  7 M (0.83333333 0.16666667)  
#>      4) V10>=0.11635 33  0 M (1.00000000 0.00000000) *
#>      5) V10< 0.11635 9  2 R (0.22222222 0.77777778) *
#>    3) V37>=0.1872 97 30 R (0.30927835 0.69072165)  
#>      6) V54>=0.016 19  3 M (0.84210526 0.15789474) *
#>      7) V54< 0.016 78 14 R (0.17948718 0.82051282)  
#>       14) V40>=0.3336 26 12 R (0.46153846 0.53846154)  
#>         28) V33< 0.53175 12  1 M (0.91666667 0.08333333) *
#>         29) V33>=0.53175 14  1 R (0.07142857 0.92857143) *
#>       15) V40< 0.3336 52  2 R (0.03846154 0.96153846) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V36< 0.4555 77 27 M (0.64935065 0.35064935)  
#>      4) V42>=0.22105 52  9 M (0.82692308 0.17307692)  
#>        8) V17>=0.1678 45  2 M (0.95555556 0.04444444) *
#>        9) V17< 0.1678 7  0 R (0.00000000 1.00000000) *
#>      5) V42< 0.22105 25  7 R (0.28000000 0.72000000)  
#>       10) V32>=0.32375 8  2 M (0.75000000 0.25000000) *
#>       11) V32< 0.32375 17  1 R (0.05882353 0.94117647) *
#>    3) V36>=0.4555 62 13 R (0.20967742 0.79032258)  
#>      6) V59< 0.00495 8  1 M (0.87500000 0.12500000) *
#>      7) V59>=0.00495 54  6 R (0.11111111 0.88888889)  
#>       14) V57>=0.0185 7  2 M (0.71428571 0.28571429) *
#>       15) V57< 0.0185 47  1 R (0.02127660 0.97872340) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V11>=0.19795 90 28 M (0.68888889 0.31111111)  
#>      4) V47>=0.06675 69 11 M (0.84057971 0.15942029)  
#>        8) V39>=0.10325 60  4 M (0.93333333 0.06666667) *
#>        9) V39< 0.10325 9  2 R (0.22222222 0.77777778) *
#>      5) V47< 0.06675 21  4 R (0.19047619 0.80952381)  
#>       10) V44< 0.12605 7  3 M (0.57142857 0.42857143) *
#>       11) V44>=0.12605 14  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.19795 49  6 R (0.12244898 0.87755102)  
#>      6) V21>=0.8695 8  3 M (0.62500000 0.37500000) *
#>      7) V21< 0.8695 41  1 R (0.02439024 0.97560976) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V45>=0.1728 60 19 M (0.68333333 0.31666667)  
#>      4) V17>=0.26765 34  2 M (0.94117647 0.05882353) *
#>      5) V17< 0.26765 26  9 R (0.34615385 0.65384615)  
#>       10) V29>=0.76855 9  1 M (0.88888889 0.11111111) *
#>       11) V29< 0.76855 17  1 R (0.05882353 0.94117647) *
#>    3) V45< 0.1728 79 18 R (0.22784810 0.77215190)  
#>      6) V48>=0.07585 21  9 M (0.57142857 0.42857143)  
#>       12) V37< 0.44055 13  1 M (0.92307692 0.07692308) *
#>       13) V37>=0.44055 8  0 R (0.00000000 1.00000000) *
#>      7) V48< 0.07585 58  6 R (0.10344828 0.89655172)  
#>       14) V6< 0.0613 7  3 M (0.57142857 0.42857143) *
#>       15) V6>=0.0613 51  2 R (0.03921569 0.96078431) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.4748201 0.5251799)  
#>    2) V9>=0.10235 99 36 M (0.6363636 0.3636364)  
#>      4) V17< 0.715 71 14 M (0.8028169 0.1971831)  
#>        8) V14>=0.19335 46  0 M (1.0000000 0.0000000) *
#>        9) V14< 0.19335 25 11 R (0.4400000 0.5600000)  
#>         18) V48>=0.08555 10  1 M (0.9000000 0.1000000) *
#>         19) V48< 0.08555 15  2 R (0.1333333 0.8666667) *
#>      5) V17>=0.715 28  6 R (0.2142857 0.7857143)  
#>       10) V58>=0.00975 7  1 M (0.8571429 0.1428571) *
#>       11) V58< 0.00975 21  0 R (0.0000000 1.0000000) *
#>    3) V9< 0.10235 40  3 R (0.0750000 0.9250000) *
#> 
#> 
#> $weights
#>  [1] 0.7346456 0.9633510 0.6881667 0.4065586 0.5451998 0.5607876 0.7108614
#>  [8] 0.6225701 0.7203105 0.8389955
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.1045058 4.6869409
#>   [2,] 1.6864040 5.1050427
#>   [3,] 1.3699096 5.4215371
#>   [4,] 2.4746357 4.3168110
#>   [5,] 2.6132769 4.1781698
#>   [6,] 3.5196419 3.2718048
#>   [7,] 2.5989418 4.1925049
#>   [8,] 1.2560611 5.5353856
#>   [9,] 1.4311718 5.3602749
#>  [10,] 1.7494392 5.0420075
#>  [11,] 2.9675021 3.8239446
#>  [12,] 2.7414110 4.0500357
#>  [13,] 1.9024156 4.8890311
#>  [14,] 1.2455541 5.5458926
#>  [15,] 0.0000000 6.7914467
#>  [16,] 1.9835999 4.8078468
#>  [17,] 2.1658175 4.6256292
#>  [18,] 2.0580763 4.7333704
#>  [19,] 0.7203105 6.0711362
#>  [20,] 0.9517584 5.8396883
#>  [21,] 1.2655103 5.5259364
#>  [22,] 2.2624952 4.5289515
#>  [23,] 2.6331495 4.1582972
#>  [24,] 1.1268691 5.6645776
#>  [25,] 1.8168487 4.9745980
#>  [26,] 2.9210231 3.8704236
#>  [27,] 1.5125460 5.2789007
#>  [28,] 2.0950566 4.6963901
#>  [29,] 1.3572158 5.4342309
#>  [30,] 1.6626198 5.1288269
#>  [31,] 0.0000000 6.7914467
#>  [32,] 0.0000000 6.7914467
#>  [33,] 1.2655103 5.5259364
#>  [34,] 0.7203105 6.0711362
#>  [35,] 1.3572158 5.4342309
#>  [36,] 2.6188638 4.1725829
#>  [37,] 2.5050154 4.2864313
#>  [38,] 2.1497323 4.6417144
#>  [39,] 0.4065586 6.3848881
#>  [40,] 1.8150357 4.9764109
#>  [41,] 1.8055866 4.9858601
#>  [42,] 0.5607876 6.2306591
#>  [43,] 1.4311718 5.3602749
#>  [44,] 2.1658175 4.6256292
#>  [45,] 0.7108614 6.0805853
#>  [46,] 0.7108614 6.0805853
#>  [47,] 0.0000000 6.7914467
#>  [48,] 0.7108614 6.0805853
#>  [49,] 0.0000000 6.7914467
#>  [50,] 0.9517584 5.8396883
#>  [51,] 0.0000000 6.7914467
#>  [52,] 2.3602355 4.4312112
#>  [53,] 0.0000000 6.7914467
#>  [54,] 0.7203105 6.0711362
#>  [55,] 0.9633510 5.8280957
#>  [56,] 0.0000000 6.7914467
#>  [57,] 0.5451998 6.2462469
#>  [58,] 2.0215981 4.7698486
#>  [59,] 2.0215981 4.7698486
#>  [60,] 2.5667979 4.2246488
#>  [61,] 1.3107368 5.4807099
#>  [62,] 1.8715244 4.9199223
#>  [63,] 2.3718281 4.4196186
#>  [64,] 2.0157437 4.7757030
#>  [65,] 1.4549561 5.3364906
#>  [66,] 1.9801997 4.8112470
#>  [67,] 2.1106444 4.6808023
#>  [68,] 1.8063417 4.9851050
#>  [69,] 1.5125460 5.2789007
#>  [70,] 1.8063417 4.9851050
#>  [71,] 4.4708799 2.3205668
#>  [72,] 5.4960135 1.2954332
#>  [73,] 4.2782389 2.5132078
#>  [74,] 5.1172344 1.6742123
#>  [75,] 4.9891002 1.8023465
#>  [76,] 5.0934501 1.6979966
#>  [77,] 4.6868915 2.1045552
#>  [78,] 4.8112470 1.9801997
#>  [79,] 5.1077852 1.6836615
#>  [80,] 5.4485661 1.3428806
#>  [81,] 6.7914467 0.0000000
#>  [82,] 4.4439004 2.3475463
#>  [83,] 6.0805853 0.7108614
#>  [84,] 6.7914467 0.0000000
#>  [85,] 5.5353856 1.2560611
#>  [86,] 4.3425787 2.4488680
#>  [87,] 6.2306591 0.5607876
#>  [88,] 6.2306591 0.5607876
#>  [89,] 5.6854593 1.1059874
#>  [90,] 6.0711362 0.7203105
#>  [91,] 5.5353856 1.2560611
#>  [92,] 5.1288269 1.6626198
#>  [93,] 4.8150751 1.9763716
#>  [94,] 6.2306591 0.5607876
#>  [95,] 6.2306591 0.5607876
#>  [96,] 6.3848881 0.4065586
#>  [97,] 4.9537163 1.8377304
#>  [98,] 5.8280957 0.9633510
#>  [99,] 4.8890311 1.9024156
#> [100,] 5.8241005 0.9673462
#> [101,] 5.3924187 1.3990280
#> [102,] 5.6967214 1.0947253
#> [103,] 4.8734434 1.9180033
#> [104,] 4.6721082 2.1193385
#> [105,] 6.7914467 0.0000000
#> [106,] 4.7377047 2.0537420
#> [107,] 5.2673081 1.5241386
#> [108,] 5.3459397 1.4455070
#> [109,] 5.6502424 1.1412042
#> [110,] 4.6726058 2.1188409
#> [111,] 6.2306591 0.5607876
#> [112,] 5.8241005 0.9673462
#> [113,] 4.9745980 1.8168487
#> [114,] 5.5580803 1.2333664
#> [115,] 6.2306591 0.5607876
#> [116,] 5.6740267 1.1174200
#> [117,] 4.7233696 2.0680771
#> [118,] 3.5005102 3.2909365
#> [119,] 4.1582972 2.6331495
#> [120,] 5.4072514 1.3841953
#> [121,] 3.7080022 3.0834445
#> [122,] 4.5469977 2.2444490
#> [123,] 4.8377698 1.9536769
#> [124,] 6.2462469 0.5451998
#> [125,] 6.2462469 0.5451998
#> [126,] 5.2789007 1.5125460
#> [127,] 6.7914467 0.0000000
#> [128,] 5.8396883 0.9517584
#> [129,] 6.2462469 0.5451998
#> [130,] 5.2015304 1.5899163
#> [131,] 6.1032800 0.6881667
#> [132,] 6.7914467 0.0000000
#> [133,] 6.7914467 0.0000000
#> [134,] 6.1032800 0.6881667
#> [135,] 6.1032800 0.6881667
#> [136,] 4.7698486 2.0215981
#> [137,] 5.6967214 1.0947253
#> [138,] 6.1032800 0.6881667
#> [139,] 4.0809269 2.7105198
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.30987591 0.69012409
#>   [2,] 0.24831293 0.75168707
#>   [3,] 0.20171101 0.79828899
#>   [4,] 0.36437534 0.63562466
#>   [5,] 0.38478943 0.61521057
#>   [6,] 0.51824626 0.48175374
#>   [7,] 0.38267866 0.61732134
#>   [8,] 0.18494751 0.81505249
#>   [9,] 0.21073151 0.78926849
#>  [10,] 0.25759449 0.74240551
#>  [11,] 0.43694697 0.56305303
#>  [12,] 0.40365642 0.59634358
#>  [13,] 0.28011934 0.71988066
#>  [14,] 0.18340041 0.81659959
#>  [15,] 0.00000000 1.00000000
#>  [16,] 0.29207324 0.70792676
#>  [17,] 0.31890370 0.68109630
#>  [18,] 0.30303945 0.69696055
#>  [19,] 0.10606142 0.89393858
#>  [20,] 0.14014074 0.85985926
#>  [21,] 0.18633884 0.81366116
#>  [22,] 0.33313891 0.66686109
#>  [23,] 0.38771555 0.61228445
#>  [24,] 0.16592475 0.83407525
#>  [25,] 0.26752013 0.73247987
#>  [26,] 0.43010322 0.56989678
#>  [27,] 0.22271337 0.77728663
#>  [28,] 0.30848459 0.69151541
#>  [29,] 0.19984192 0.80015808
#>  [30,] 0.24481084 0.75518916
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.00000000 1.00000000
#>  [33,] 0.18633884 0.81366116
#>  [34,] 0.10606142 0.89393858
#>  [35,] 0.19984192 0.80015808
#>  [36,] 0.38561207 0.61438793
#>  [37,] 0.36884857 0.63115143
#>  [38,] 0.31653525 0.68346475
#>  [39,] 0.05986333 0.94013667
#>  [40,] 0.26725318 0.73274682
#>  [41,] 0.26586186 0.73413814
#>  [42,] 0.08257262 0.91742738
#>  [43,] 0.21073151 0.78926849
#>  [44,] 0.31890370 0.68109630
#>  [45,] 0.10467009 0.89532991
#>  [46,] 0.10467009 0.89532991
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.10467009 0.89532991
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.14014074 0.85985926
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.34753060 0.65246940
#>  [53,] 0.00000000 1.00000000
#>  [54,] 0.10606142 0.89393858
#>  [55,] 0.14184768 0.85815232
#>  [56,] 0.00000000 1.00000000
#>  [57,] 0.08027742 0.91972258
#>  [58,] 0.29766826 0.70233174
#>  [59,] 0.29766826 0.70233174
#>  [60,] 0.37794568 0.62205432
#>  [61,] 0.19299817 0.80700183
#>  [62,] 0.27557079 0.72442921
#>  [63,] 0.34923754 0.65076246
#>  [64,] 0.29680623 0.70319377
#>  [65,] 0.21423361 0.78576639
#>  [66,] 0.29157259 0.70842741
#>  [67,] 0.31077979 0.68922021
#>  [68,] 0.26597303 0.73402697
#>  [69,] 0.22271337 0.77728663
#>  [70,] 0.26597303 0.73402697
#>  [71,] 0.65831039 0.34168961
#>  [72,] 0.80925519 0.19074481
#>  [73,] 0.62994514 0.37005486
#>  [74,] 0.75348222 0.24651778
#>  [75,] 0.73461524 0.26538476
#>  [76,] 0.74998013 0.25001987
#>  [77,] 0.69011680 0.30988320
#>  [78,] 0.70842741 0.29157259
#>  [79,] 0.75209090 0.24790910
#>  [80,] 0.80226884 0.19773116
#>  [81,] 1.00000000 0.00000000
#>  [82,] 0.65433782 0.34566218
#>  [83,] 0.89532991 0.10467009
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.81505249 0.18494751
#>  [86,] 0.63941881 0.36058119
#>  [87,] 0.91742738 0.08257262
#>  [88,] 0.91742738 0.08257262
#>  [89,] 0.83714996 0.16285004
#>  [90,] 0.89393858 0.10606142
#>  [91,] 0.81505249 0.18494751
#>  [92,] 0.75518916 0.24481084
#>  [93,] 0.70899107 0.29100893
#>  [94,] 0.91742738 0.08257262
#>  [95,] 0.91742738 0.08257262
#>  [96,] 0.94013667 0.05986333
#>  [97,] 0.72940516 0.27059484
#>  [98,] 0.85815232 0.14184768
#>  [99,] 0.71988066 0.28011934
#> [100,] 0.85756405 0.14243595
#> [101,] 0.79400147 0.20599853
#> [102,] 0.83880824 0.16119176
#> [103,] 0.71758545 0.28241455
#> [104,] 0.68794005 0.31205995
#> [105,] 1.00000000 0.00000000
#> [106,] 0.69759875 0.30240125
#> [107,] 0.77557969 0.22442031
#> [108,] 0.78715772 0.21284228
#> [109,] 0.83196448 0.16803552
#> [110,] 0.68801332 0.31198668
#> [111,] 0.91742738 0.08257262
#> [112,] 0.85756405 0.14243595
#> [113,] 0.73247987 0.26752013
#> [114,] 0.81839415 0.18160585
#> [115,] 0.91742738 0.08257262
#> [116,] 0.83546658 0.16453342
#> [117,] 0.69548798 0.30451202
#> [118,] 0.51542923 0.48457077
#> [119,] 0.61228445 0.38771555
#> [120,] 0.79618551 0.20381449
#> [121,] 0.54598119 0.45401881
#> [122,] 0.66951827 0.33048173
#> [123,] 0.71233273 0.28766727
#> [124,] 0.91972258 0.08027742
#> [125,] 0.91972258 0.08027742
#> [126,] 0.77728663 0.22271337
#> [127,] 1.00000000 0.00000000
#> [128,] 0.85985926 0.14014074
#> [129,] 0.91972258 0.08027742
#> [130,] 0.76589431 0.23410569
#> [131,] 0.89867157 0.10132843
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 0.89867157 0.10132843
#> [135,] 0.89867157 0.10132843
#> [136,] 0.70233174 0.29766826
#> [137,] 0.83880824 0.16119176
#> [138,] 0.89867157 0.10132843
#> [139,] 0.60089213 0.39910787
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "M" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  1.3811406 10.1877648  2.2729732  0.0000000  4.2117100  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  2.9143037  7.7866992  0.0000000  0.0000000  0.0000000  0.0000000  3.4986218 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  1.3117694  0.5142661  0.0000000  0.0000000  0.5005148  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  1.6814490  0.0000000  0.0000000  0.0000000  1.0636114  1.4904374  0.6127836 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  2.7172767  6.9850103  0.0000000  1.4184994  1.9513324  1.0017064 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  3.6981614  0.0000000  0.5461856  5.1307525  5.4231041  6.5319449 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.1037973  0.0000000  0.9729561  0.0000000  1.5261370  0.0000000  0.5353675 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  2.6918815  1.9225450  0.0000000  1.1975295  1.8631449  1.6640204  0.7230228 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  8.9675794 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 69 70 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11         V9        V17        V37        V47        V46        V45 
#> 10.1877648  8.9675794  7.7866992  6.9850103  6.5319449  5.4231041  5.1307525 
#>        V14        V42        V21        V48        V16        V36        V54 
#>  4.2117100  3.6981614  3.4986218  3.1037973  2.9143037  2.7172767  2.6918815 
#>        V12         V4        V55        V58        V29        V59        V51 
#>  2.2729732  1.9513324  1.9225450  1.8631449  1.6814490  1.6640204  1.5261370 
#>        V33        V39        V10        V23        V57        V32        V40 
#>  1.4904374  1.4184994  1.3811406  1.3117694  1.1975295  1.0636114  1.0017064 
#>         V5         V6        V34        V44        V53        V24        V27 
#>  0.9729561  0.7230228  0.6127836  0.5461856  0.5353675  0.5142661  0.5005148 
#>         V1        V13        V15        V18        V19         V2        V20 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V22        V25        V26        V28         V3        V30        V31 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V38        V41        V43        V49        V50        V52 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V56        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058