Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V12>=0.21345 83 14 M (0.83132530 0.16867470)  
#>      4) V15< 0.67555 73  7 M (0.90410959 0.09589041) *
#>      5) V15>=0.67555 10  3 R (0.30000000 0.70000000) *
#>    3) V12< 0.21345 56 12 R (0.21428571 0.78571429)  
#>      6) V1>=0.02065 21  9 M (0.57142857 0.42857143)  
#>       12) V26< 0.6282 11  0 M (1.00000000 0.00000000) *
#>       13) V26>=0.6282 10  1 R (0.10000000 0.90000000) *
#>      7) V1< 0.02065 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.1791 89 23 M (0.74157303 0.25842697)  
#>      4) V48>=0.0755 54  4 M (0.92592593 0.07407407) *
#>      5) V48< 0.0755 35 16 R (0.45714286 0.54285714)  
#>       10) V5< 0.04495 13  1 M (0.92307692 0.07692308) *
#>       11) V5>=0.04495 22  4 R (0.18181818 0.81818182) *
#>    3) V11< 0.1791 50  9 R (0.18000000 0.82000000)  
#>      6) V1>=0.0324 7  1 M (0.85714286 0.14285714) *
#>      7) V1< 0.0324 43  3 R (0.06976744 0.93023256) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 R (0.3956835 0.6043165)  
#>    2) V5>=0.0593 76 31 M (0.5921053 0.4078947)  
#>      4) V49>=0.04605 40  7 M (0.8250000 0.1750000)  
#>        8) V14< 0.5179 32  1 M (0.9687500 0.0312500) *
#>        9) V14>=0.5179 8  2 R (0.2500000 0.7500000) *
#>      5) V49< 0.04605 36 12 R (0.3333333 0.6666667)  
#>       10) V54>=0.01675 8  0 M (1.0000000 0.0000000) *
#>       11) V54< 0.01675 28  4 R (0.1428571 0.8571429)  
#>         22) V29>=0.7915 7  3 M (0.5714286 0.4285714) *
#>         23) V29< 0.7915 21  0 R (0.0000000 1.0000000) *
#>    3) V5< 0.0593 63 10 R (0.1587302 0.8412698) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.1238 102 29 M (0.71568627 0.28431373)  
#>      4) V7< 0.2313 92 19 M (0.79347826 0.20652174)  
#>        8) V52>=0.00675 80 10 M (0.87500000 0.12500000)  
#>         16) V37< 0.4645 66  2 M (0.96969697 0.03030303) *
#>         17) V37>=0.4645 14  6 R (0.42857143 0.57142857) *
#>        9) V52< 0.00675 12  3 R (0.25000000 0.75000000) *
#>      5) V7>=0.2313 10  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1238 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 60 M (0.56834532 0.43165468)  
#>   2) V11>=0.19795 88 19 M (0.78409091 0.21590909)  
#>     4) V39>=0.15985 76 11 M (0.85526316 0.14473684)  
#>       8) V52>=0.00465 69  6 M (0.91304348 0.08695652) *
#>       9) V52< 0.00465 7  2 R (0.28571429 0.71428571) *
#>     5) V39< 0.15985 12  4 R (0.33333333 0.66666667) *
#>   3) V11< 0.19795 51 10 R (0.19607843 0.80392157)  
#>     6) V59>=0.00775 13  4 M (0.69230769 0.30769231) *
#>     7) V59< 0.00775 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V23>=0.78505 51  9 M (0.82352941 0.17647059)  
#>      4) V4< 0.1043 43  3 M (0.93023256 0.06976744) *
#>      5) V4>=0.1043 8  2 R (0.25000000 0.75000000) *
#>    3) V23< 0.78505 88 29 R (0.32954545 0.67045455)  
#>      6) V49>=0.08415 14  0 M (1.00000000 0.00000000) *
#>      7) V49< 0.08415 74 15 R (0.20270270 0.79729730)  
#>       14) V36< 0.14435 15  4 M (0.73333333 0.26666667) *
#>       15) V36>=0.14435 59  4 R (0.06779661 0.93220339) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 M (0.54676259 0.45323741)  
#>   2) V46>=0.0743 87 24 M (0.72413793 0.27586207)  
#>     4) V8>=0.0742 73 11 M (0.84931507 0.15068493)  
#>       8) V11>=0.1117 65  5 M (0.92307692 0.07692308) *
#>       9) V11< 0.1117 8  2 R (0.25000000 0.75000000) *
#>     5) V8< 0.0742 14  1 R (0.07142857 0.92857143) *
#>   3) V46< 0.0743 52 13 R (0.25000000 0.75000000)  
#>     6) V28< 0.54175 15  3 M (0.80000000 0.20000000) *
#>     7) V28>=0.54175 37  1 R (0.02702703 0.97297297) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V4>=0.05945 54  9 M (0.83333333 0.16666667)  
#>      4) V37< 0.5049 41  1 M (0.97560976 0.02439024) *
#>      5) V37>=0.5049 13  5 R (0.38461538 0.61538462) *
#>    3) V4< 0.05945 85 25 R (0.29411765 0.70588235)  
#>      6) V45>=0.314 13  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.314 72 12 R (0.16666667 0.83333333)  
#>       14) V25>=0.9245 7  1 M (0.85714286 0.14285714) *
#>       15) V25< 0.9245 65  6 R (0.09230769 0.90769231) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V43>=0.22245 73 19 M (0.73972603 0.26027397)  
#>      4) V40< 0.61335 62  8 M (0.87096774 0.12903226)  
#>        8) V5>=0.03745 55  4 M (0.92727273 0.07272727) *
#>        9) V5< 0.03745 7  3 R (0.42857143 0.57142857) *
#>      5) V40>=0.61335 11  0 R (0.00000000 1.00000000) *
#>    3) V43< 0.22245 66 20 R (0.30303030 0.69696970)  
#>      6) V27< 0.22585 9  0 M (1.00000000 0.00000000) *
#>      7) V27>=0.22585 57 11 R (0.19298246 0.80701754)  
#>       14) V49>=0.04705 11  3 M (0.72727273 0.27272727) *
#>       15) V49< 0.04705 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V4>=0.0606 53 16 M (0.69811321 0.30188679)  
#>      4) V53>=0.00405 44  7 M (0.84090909 0.15909091)  
#>        8) V7< 0.2101 37  1 M (0.97297297 0.02702703) *
#>        9) V7>=0.2101 7  1 R (0.14285714 0.85714286) *
#>      5) V53< 0.00405 9  0 R (0.00000000 1.00000000) *
#>    3) V4< 0.0606 86 26 R (0.30232558 0.69767442)  
#>      6) V35< 0.1318 14  1 M (0.92857143 0.07142857) *
#>      7) V35>=0.1318 72 13 R (0.18055556 0.81944444)  
#>       14) V30>=0.8334 10  2 M (0.80000000 0.20000000) *
#>       15) V30< 0.8334 62  5 R (0.08064516 0.91935484) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.6054686 0.3936344 0.6868210 0.6361118 0.6916808 0.6088189
#>  [8] 0.6226057 0.7248635 0.6690271
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.9697941 4.3805686
#>   [2,] 2.5543574 3.7960054
#>   [3,] 1.3474428 5.0029200
#>   [4,] 0.6226057 5.7277570
#>   [5,] 0.7113310 5.6390317
#>   [6,] 1.2142875 5.1360752
#>   [7,] 1.7875250 4.5628378
#>   [8,] 2.7259445 3.6244182
#>   [9,] 1.1184979 5.2318648
#>  [10,] 1.3607079 4.9896549
#>  [11,] 0.6690271 5.6813357
#>  [12,] 1.3004997 5.0498630
#>  [13,] 1.7137844 4.6365783
#>  [14,] 2.7140405 3.6363223
#>  [15,] 2.4236367 3.9267260
#>  [16,] 2.5788674 3.7714953
#>  [17,] 1.9562617 4.3941010
#>  [18,] 2.5617303 3.7886324
#>  [19,] 2.9528516 3.3975111
#>  [20,] 0.6690271 5.6813357
#>  [21,] 0.6690271 5.6813357
#>  [22,] 1.2314246 5.1189381
#>  [23,] 2.4143921 3.9359706
#>  [24,] 2.6777748 3.6725879
#>  [25,] 1.3167996 5.0335631
#>  [26,] 0.6868210 5.6635417
#>  [27,] 0.0000000 6.3503627
#>  [28,] 2.0084804 4.3418824
#>  [29,] 2.0450134 4.3053493
#>  [30,] 2.0146135 4.3357492
#>  [31,] 2.1033652 4.2469975
#>  [32,] 1.9919598 4.3584029
#>  [33,] 0.6690271 5.6813357
#>  [34,] 2.0723062 4.2780565
#>  [35,] 1.2415804 5.1087824
#>  [36,] 0.0000000 6.3503627
#>  [37,] 0.6690271 5.6813357
#>  [38,] 0.0000000 6.3503627
#>  [39,] 0.7248635 5.6254993
#>  [40,] 2.0205034 4.3298593
#>  [41,] 2.5059203 3.8444424
#>  [42,] 1.0024534 5.3479093
#>  [43,] 1.3277925 5.0225702
#>  [44,] 1.0853152 5.2650475
#>  [45,] 0.0000000 6.3503627
#>  [46,] 0.6361118 5.7142510
#>  [47,] 1.3474428 5.0029200
#>  [48,] 2.6253152 3.7250476
#>  [49,] 0.0000000 6.3503627
#>  [50,] 0.6088189 5.7415438
#>  [51,] 0.6916808 5.6586820
#>  [52,] 1.3142865 5.0360763
#>  [53,] 0.6916808 5.6586820
#>  [54,] 1.2971494 5.0532134
#>  [55,] 2.0855713 4.2647914
#>  [56,] 1.9695268 4.3808359
#>  [57,] 1.3474428 5.0029200
#>  [58,] 2.6259720 3.7243907
#>  [59,] 2.6172993 3.7330634
#>  [60,] 1.9391510 4.4112117
#>  [61,] 2.3327854 4.0175773
#>  [62,] 1.6385651 4.7117976
#>  [63,] 1.4030118 4.9473509
#>  [64,] 2.0164698 4.3338929
#>  [65,] 0.6361118 5.7142510
#>  [66,] 5.0498630 1.3004997
#>  [67,] 5.6586820 0.6916808
#>  [68,] 5.9567283 0.3936344
#>  [69,] 4.5205338 1.8298289
#>  [70,] 5.6586820 0.6916808
#>  [71,] 3.7615806 2.5887821
#>  [72,] 3.3679461 2.9824166
#>  [73,] 3.9347155 2.4156473
#>  [74,] 4.3630420 1.9873207
#>  [75,] 4.3467421 2.0036206
#>  [76,] 3.9531077 2.3972550
#>  [77,] 5.6586820 0.6916808
#>  [78,] 5.7448941 0.6054686
#>  [79,] 4.0596269 2.2907358
#>  [80,] 5.0587300 1.2916328
#>  [81,] 4.6337955 1.7165672
#>  [82,] 5.7277570 0.6226057
#>  [83,] 6.3503627 0.0000000
#>  [84,] 5.0758671 1.2744957
#>  [85,] 4.5450438 1.8053189
#>  [86,] 6.3503627 0.0000000
#>  [87,] 5.0274300 1.3229327
#>  [88,] 6.3503627 0.0000000
#>  [89,] 4.3510036 1.9993591
#>  [90,] 4.3510036 1.9993591
#>  [91,] 4.6227916 1.7275711
#>  [92,] 3.2289011 3.1214617
#>  [93,] 4.6399287 1.7104340
#>  [94,] 4.6399287 1.7104340
#>  [95,] 4.6008802 1.7494825
#>  [96,] 5.0409361 1.3094267
#>  [97,] 4.3492553 2.0011075
#>  [98,] 5.2318648 1.1184979
#>  [99,] 3.6802282 2.6701345
#> [100,] 4.6230459 1.7273168
#> [101,] 6.3503627 0.0000000
#> [102,] 5.6390317 0.7113310
#> [103,] 3.9313651 2.4189976
#> [104,] 5.0225702 1.3277925
#> [105,] 5.3512597 0.9991030
#> [106,] 5.0274300 1.3229327
#> [107,] 4.0925422 2.2578205
#> [108,] 3.9402321 2.4101307
#> [109,] 4.4532614 1.8971014
#> [110,] 5.0409361 1.3094267
#> [111,] 5.6635417 0.6868210
#> [112,] 5.0409361 1.3094267
#> [113,] 5.2699073 1.0804554
#> [114,] 5.2877012 1.0626615
#> [115,] 5.2318648 1.1184979
#> [116,] 3.3468019 3.0035608
#> [117,] 3.9920613 2.3583015
#> [118,] 4.3160726 2.0342901
#> [119,] 5.0532134 1.2971494
#> [120,] 4.6644387 1.6859240
#> [121,] 4.5450438 1.8053189
#> [122,] 4.4272573 1.9231054
#> [123,] 6.3503627 0.0000000
#> [124,] 6.3503627 0.0000000
#> [125,] 5.6813357 0.6690271
#> [126,] 5.9567283 0.3936344
#> [127,] 6.3503627 0.0000000
#> [128,] 5.0916453 1.2587174
#> [129,] 5.0587300 1.2916328
#> [130,] 3.9402321 2.4101307
#> [131,] 5.7142510 0.6361118
#> [132,] 5.0452239 1.3051388
#> [133,] 3.9267260 2.4236367
#> [134,] 5.7415438 0.6088189
#> [135,] 5.3341226 1.0162401
#> [136,] 5.0547228 1.2956399
#> [137,] 5.2650475 1.0853152
#> [138,] 5.3206165 1.0297462
#> [139,] 4.5957531 1.7546097
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.31018608 0.68981392
#>   [2,] 0.40223803 0.59776197
#>   [3,] 0.21218359 0.78781641
#>   [4,] 0.09804254 0.90195746
#>   [5,] 0.11201423 0.88798577
#>   [6,] 0.19121546 0.80878454
#>   [7,] 0.28148392 0.71851608
#>   [8,] 0.42925808 0.57074192
#>   [9,] 0.17613134 0.82386866
#>  [10,] 0.21427246 0.78572754
#>  [11,] 0.10535258 0.89464742
#>  [12,] 0.20479141 0.79520859
#>  [13,] 0.26987189 0.73012811
#>  [14,] 0.42738354 0.57261646
#>  [15,] 0.38165327 0.61834673
#>  [16,] 0.40609765 0.59390235
#>  [17,] 0.30805511 0.69194489
#>  [18,] 0.40339905 0.59660095
#>  [19,] 0.46498944 0.53501056
#>  [20,] 0.10535258 0.89464742
#>  [21,] 0.10535258 0.89464742
#>  [22,] 0.19391406 0.80608594
#>  [23,] 0.38019751 0.61980249
#>  [24,] 0.42167273 0.57832727
#>  [25,] 0.20735817 0.79264183
#>  [26,] 0.10815461 0.89184539
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.31627806 0.68372194
#>  [29,] 0.32203096 0.67796904
#>  [30,] 0.31724385 0.68275615
#>  [31,] 0.33121970 0.66878030
#>  [32,] 0.31367654 0.68632346
#>  [33,] 0.10535258 0.89464742
#>  [34,] 0.32632879 0.67367121
#>  [35,] 0.19551330 0.80448670
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.10535258 0.89464742
#>  [38,] 0.00000000 1.00000000
#>  [39,] 0.11414520 0.88585480
#>  [40,] 0.31817133 0.68182867
#>  [41,] 0.39461058 0.60538942
#>  [42,] 0.15785766 0.84214234
#>  [43,] 0.20908924 0.79091076
#>  [44,] 0.17090602 0.82909398
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.10016936 0.89983064
#>  [47,] 0.21218359 0.78781641
#>  [48,] 0.41341184 0.58658816
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.09587152 0.90412848
#>  [51,] 0.10891989 0.89108011
#>  [52,] 0.20696242 0.79303758
#>  [53,] 0.10891989 0.89108011
#>  [54,] 0.20426382 0.79573618
#>  [55,] 0.32841767 0.67158233
#>  [56,] 0.31014398 0.68985602
#>  [57,] 0.21218359 0.78781641
#>  [58,] 0.41351527 0.58648473
#>  [59,] 0.41214958 0.58785042
#>  [60,] 0.30536067 0.69463933
#>  [61,] 0.36734680 0.63265320
#>  [62,] 0.25802701 0.74197299
#>  [63,] 0.22093412 0.77906588
#>  [64,] 0.31753617 0.68246383
#>  [65,] 0.10016936 0.89983064
#>  [66,] 0.79520859 0.20479141
#>  [67,] 0.89108011 0.10891989
#>  [68,] 0.93801386 0.06198614
#>  [69,] 0.71185443 0.28814557
#>  [70,] 0.89108011 0.10891989
#>  [71,] 0.59234106 0.40765894
#>  [72,] 0.53035492 0.46964508
#>  [73,] 0.61960484 0.38039516
#>  [74,] 0.68705398 0.31294602
#>  [75,] 0.68448722 0.31551278
#>  [76,] 0.62250109 0.37749891
#>  [77,] 0.89108011 0.10891989
#>  [78,] 0.90465606 0.09534394
#>  [79,] 0.63927481 0.36072519
#>  [80,] 0.79660488 0.20339512
#>  [81,] 0.72968990 0.27031010
#>  [82,] 0.90195746 0.09804254
#>  [83,] 1.00000000 0.00000000
#>  [84,] 0.79930348 0.20069652
#>  [85,] 0.71571405 0.28428595
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.79167603 0.20832397
#>  [88,] 1.00000000 0.00000000
#>  [89,] 0.68515828 0.31484172
#>  [90,] 0.68515828 0.31484172
#>  [91,] 0.72795709 0.27204291
#>  [92,] 0.50845931 0.49154069
#>  [93,] 0.73065569 0.26934431
#>  [94,] 0.73065569 0.26934431
#>  [95,] 0.72450668 0.27549332
#>  [96,] 0.79380285 0.20619715
#>  [97,] 0.68488297 0.31511703
#>  [98,] 0.82386866 0.17613134
#>  [99,] 0.57953039 0.42046961
#> [100,] 0.72799714 0.27200286
#> [101,] 1.00000000 0.00000000
#> [102,] 0.88798577 0.11201423
#> [103,] 0.61907725 0.38092275
#> [104,] 0.79091076 0.20908924
#> [105,] 0.84266993 0.15733007
#> [106,] 0.79167603 0.20832397
#> [107,] 0.64445803 0.35554197
#> [108,] 0.62047354 0.37952646
#> [109,] 0.70126095 0.29873905
#> [110,] 0.79380285 0.20619715
#> [111,] 0.89184539 0.10815461
#> [112,] 0.79380285 0.20619715
#> [113,] 0.82985926 0.17014074
#> [114,] 0.83266129 0.16733871
#> [115,] 0.82386866 0.17613134
#> [116,] 0.52702531 0.47297469
#> [117,] 0.62863516 0.37136484
#> [118,] 0.67965765 0.32034235
#> [119,] 0.79573618 0.20426382
#> [120,] 0.73451532 0.26548468
#> [121,] 0.71571405 0.28428595
#> [122,] 0.69716605 0.30283395
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 0.89464742 0.10535258
#> [126,] 0.93801386 0.06198614
#> [127,] 1.00000000 0.00000000
#> [128,] 0.80178810 0.19821190
#> [129,] 0.79660488 0.20339512
#> [130,] 0.62047354 0.37952646
#> [131,] 0.89983064 0.10016936
#> [132,] 0.79447807 0.20552193
#> [133,] 0.61834673 0.38165327
#> [134,] 0.90412848 0.09587152
#> [135,] 0.83997133 0.16002867
#> [136,] 0.79597387 0.20402613
#> [137,] 0.82909398 0.17090602
#> [138,] 0.83784451 0.16215549
#> [139,] 0.72369930 0.27630070
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  3.5221242  0.0000000 15.4880163  6.0089838  0.0000000  0.8635365  1.5149555 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  3.6159413  0.0000000  1.5271426  2.0025440  2.4347115  2.5760740 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.4477424  0.0000000  1.9780677  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  2.9112488  2.4311583  2.9654324  0.0000000  1.1915760  7.6789446  3.4085823 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  3.1792011  0.0000000  3.1590714  2.9556790  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.8748183  6.5023051  4.2362570  0.0000000  0.0000000  2.9131687  2.3453790 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  1.1939797  0.0000000  0.0000000  0.0000000  0.0000000  1.8133432  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  4.3885459  2.8714693  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11         V4        V49        V12         V7         V5        V23 
#> 15.4880163  7.6789446  6.5023051  6.0089838  4.3885459  4.2362570  3.6159413 
#>         V1        V40        V43        V45        V37        V46        V52 
#>  3.5221242  3.4085823  3.1792011  3.1590714  2.9654324  2.9556790  2.9131687 
#>        V35         V8        V28        V27        V36        V53        V26 
#>  2.9112488  2.8714693  2.5760740  2.4347115  2.4311583  2.3453790  2.0025440 
#>        V30        V48        V59        V25        V15        V54        V39 
#>  1.9780677  1.8748183  1.8133432  1.5271426  1.5149555  1.1939797  1.1915760 
#>        V14        V29        V10        V13        V16        V17        V18 
#>  0.8635365  0.4477424  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V19         V2        V20        V21        V22        V24         V3 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V31        V32        V33        V34        V38        V41        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V44        V47        V50        V51        V55        V56        V57 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58         V6        V60         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841