Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 67 M (0.51798561 0.48201439)  
#>   2) V11>=0.19825 83 16 M (0.80722892 0.19277108)  
#>     4) V17< 0.7271 72  7 M (0.90277778 0.09722222) *
#>     5) V17>=0.7271 11  2 R (0.18181818 0.81818182) *
#>   3) V11< 0.19825 56  5 R (0.08928571 0.91071429)  
#>     6) V21>=0.8695 7  3 M (0.57142857 0.42857143) *
#>     7) V21< 0.8695 49  1 R (0.02040816 0.97959184) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 59 M (0.57553957 0.42446043)  
#>   2) V12>=0.2248 85 18 M (0.78823529 0.21176471)  
#>     4) V47>=0.06785 69  6 M (0.91304348 0.08695652)  
#>       8) V34< 0.6932 62  2 M (0.96774194 0.03225806) *
#>       9) V34>=0.6932 7  3 R (0.42857143 0.57142857) *
#>     5) V47< 0.06785 16  4 R (0.25000000 0.75000000) *
#>   3) V12< 0.2248 54 13 R (0.24074074 0.75925926)  
#>     6) V20>=0.8321 12  2 M (0.83333333 0.16666667) *
#>     7) V20< 0.8321 42  3 R (0.07142857 0.92857143) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 57 M (0.58992806 0.41007194)  
#>   2) V9>=0.1494 72  8 M (0.88888889 0.11111111) *
#>   3) V9< 0.1494 67 18 R (0.26865672 0.73134328)  
#>     6) V14>=0.3277 19  5 M (0.73684211 0.26315789) *
#>     7) V14< 0.3277 48  4 R (0.08333333 0.91666667) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V12>=0.1677 104 32 M (0.69230769 0.30769231)  
#>      4) V51>=0.01235 63  5 M (0.92063492 0.07936508) *
#>      5) V51< 0.01235 41 14 R (0.34146341 0.65853659)  
#>       10) V55< 0.0066 17  3 M (0.82352941 0.17647059) *
#>       11) V55>=0.0066 24  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.1677 35  5 R (0.14285714 0.85714286)  
#>      6) V25< 0.5373 7  3 M (0.57142857 0.42857143) *
#>      7) V25>=0.5373 28  1 R (0.03571429 0.96428571) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 55 R (0.39568345 0.60431655)  
#>    2) V4>=0.0606 43 10 M (0.76744186 0.23255814)  
#>      4) V6< 0.2331 34  3 M (0.91176471 0.08823529) *
#>      5) V6>=0.2331 9  2 R (0.22222222 0.77777778) *
#>    3) V4< 0.0606 96 22 R (0.22916667 0.77083333)  
#>      6) V23>=0.83315 24  9 M (0.62500000 0.37500000)  
#>       12) V10>=0.141 15  0 M (1.00000000 0.00000000) *
#>       13) V10< 0.141 9  0 R (0.00000000 1.00000000) *
#>      7) V23< 0.83315 72  7 R (0.09722222 0.90277778)  
#>       14) V28>=0.94915 10  3 M (0.70000000 0.30000000) *
#>       15) V28< 0.94915 62  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V9>=0.1306 91 34 M (0.62637363 0.37362637)  
#>      4) V45>=0.2726 27  0 M (1.00000000 0.00000000) *
#>      5) V45< 0.2726 64 30 R (0.46875000 0.53125000)  
#>       10) V35< 0.35925 40 10 M (0.75000000 0.25000000)  
#>         20) V21>=0.5514 27  1 M (0.96296296 0.03703704) *
#>         21) V21< 0.5514 13  4 R (0.30769231 0.69230769) *
#>       11) V35>=0.35925 24  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.1306 48  6 R (0.12500000 0.87500000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V45>=0.1953 59 17 M (0.71186441 0.28813559)  
#>      4) V32>=0.09025 52 10 M (0.80769231 0.19230769)  
#>        8) V52>=0.00305 45  3 M (0.93333333 0.06666667) *
#>        9) V52< 0.00305 7  0 R (0.00000000 1.00000000) *
#>      5) V32< 0.09025 7  0 R (0.00000000 1.00000000) *
#>    3) V45< 0.1953 80 20 R (0.25000000 0.75000000)  
#>      6) V24>=0.6678 30 14 M (0.53333333 0.46666667)  
#>       12) V48>=0.0766 13  1 M (0.92307692 0.07692308) *
#>       13) V48< 0.0766 17  4 R (0.23529412 0.76470588) *
#>      7) V24< 0.6678 50  4 R (0.08000000 0.92000000)  
#>       14) V36< 0.14875 7  3 M (0.57142857 0.42857143) *
#>       15) V36>=0.14875 43  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V16< 0.5586 106 33 M (0.68867925 0.31132075)  
#>      4) V52>=0.00765 85 15 M (0.82352941 0.17647059)  
#>        8) V18>=0.1578 76  8 M (0.89473684 0.10526316)  
#>         16) V3>=0.01965 63  1 M (0.98412698 0.01587302) *
#>         17) V3< 0.01965 13  6 R (0.46153846 0.53846154) *
#>        9) V18< 0.1578 9  2 R (0.22222222 0.77777778) *
#>      5) V52< 0.00765 21  3 R (0.14285714 0.85714286) *
#>    3) V16>=0.5586 33  5 R (0.15151515 0.84848485)  
#>      6) V60>=0.0089 7  2 M (0.71428571 0.28571429) *
#>      7) V60< 0.0089 26  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V51>=0.01255 79 27 M (0.65822785 0.34177215)  
#>      4) V36< 0.5503 56 10 M (0.82142857 0.17857143)  
#>        8) V48>=0.0482 40  0 M (1.00000000 0.00000000) *
#>        9) V48< 0.0482 16  6 R (0.37500000 0.62500000) *
#>      5) V36>=0.5503 23  6 R (0.26086957 0.73913043)  
#>       10) V45>=0.32995 7  1 M (0.85714286 0.14285714) *
#>       11) V45< 0.32995 16  0 R (0.00000000 1.00000000) *
#>    3) V51< 0.01255 60 12 R (0.20000000 0.80000000)  
#>      6) V23>=0.8272 16  6 M (0.62500000 0.37500000) *
#>      7) V23< 0.8272 44  2 R (0.04545455 0.95454545) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V36< 0.55005 103 35 M (0.66019417 0.33980583)  
#>      4) V58>=0.0031 85 19 M (0.77647059 0.22352941)  
#>        8) V30>=0.2638 74  9 M (0.87837838 0.12162162)  
#>         16) V39< 0.39395 66  2 M (0.96969697 0.03030303) *
#>         17) V39>=0.39395 8  1 R (0.12500000 0.87500000) *
#>        9) V30< 0.2638 11  1 R (0.09090909 0.90909091) *
#>      5) V58< 0.0031 18  2 R (0.11111111 0.88888889) *
#>    3) V36>=0.55005 36  5 R (0.13888889 0.86111111)  
#>      6) V36>=0.76615 7  3 M (0.57142857 0.42857143) *
#>      7) V36< 0.76615 29  1 R (0.03448276 0.96551724) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.6013343 0.5206440 0.7671926 1.0209383 0.8349688 1.0358926
#>  [8] 0.5388519 0.7416409 0.7602461
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.2808902 6.3242585
#>   [2,] 1.2878366 6.3173121
#>   [3,] 3.0690211 4.5361276
#>   [4,] 1.2878366 6.3173121
#>   [5,] 1.9054174 5.6997312
#>   [6,] 3.2114619 4.3936868
#>   [7,] 2.0407389 5.5644098
#>   [8,] 0.0000000 7.6051487
#>   [9,] 1.5274387 6.0777100
#>  [10,] 2.5865243 5.0186244
#>  [11,] 2.5774749 5.0276738
#>  [12,] 0.0000000 7.6051487
#>  [13,] 2.1228055 5.4823432
#>  [14,] 3.3517686 4.2533801
#>  [15,] 1.5274387 6.0777100
#>  [16,] 1.7625792 5.8425695
#>  [17,] 1.9054174 5.6997312
#>  [18,] 1.1401861 6.4649625
#>  [19,] 0.5388519 7.0662968
#>  [20,] 1.8636192 5.7415295
#>  [21,] 1.3685268 6.2366219
#>  [22,] 0.5388519 7.0662968
#>  [23,] 0.7602461 6.8449025
#>  [24,] 0.5388519 7.0662968
#>  [25,] 1.2808902 6.3242585
#>  [26,] 0.0000000 7.6051487
#>  [27,] 2.0662906 5.5388581
#>  [28,] 3.0835780 4.5215707
#>  [29,] 0.7834391 6.8217095
#>  [30,] 1.5415823 6.0635664
#>  [31,] 2.1101677 5.4949810
#>  [32,] 1.5747445 6.0304042
#>  [33,] 1.3847734 6.2203753
#>  [34,] 1.7961387 5.8090099
#>  [35,] 0.0000000 7.6051487
#>  [36,] 0.7602461 6.8449025
#>  [37,] 1.7961387 5.8090099
#>  [38,] 1.5018870 6.1032616
#>  [39,] 1.0209383 6.5842104
#>  [40,] 0.0000000 7.6051487
#>  [41,] 0.7602461 6.8449025
#>  [42,] 0.7602461 6.8449025
#>  [43,] 0.7602461 6.8449025
#>  [44,] 0.7602461 6.8449025
#>  [45,] 0.7602461 6.8449025
#>  [46,] 0.5388519 7.0662968
#>  [47,] 0.7602461 6.8449025
#>  [48,] 0.0000000 7.6051487
#>  [49,] 2.2197422 5.3854064
#>  [50,] 1.5250800 6.0800686
#>  [51,] 0.7834391 6.8217095
#>  [52,] 1.9569471 5.6482016
#>  [53,] 2.7171933 4.8879554
#>  [54,] 2.1578709 5.4472778
#>  [55,] 1.5415823 6.0635664
#>  [56,] 2.0953885 5.5097602
#>  [57,] 2.0953885 5.5097602
#>  [58,] 1.5565366 6.0486120
#>  [59,] 2.8315219 4.7736268
#>  [60,] 2.3237292 5.2814195
#>  [61,] 1.3040832 6.3010655
#>  [62,] 1.7881309 5.8170178
#>  [63,] 0.7671926 6.8379561
#>  [64,] 7.6051487 0.0000000
#>  [65,] 6.0614634 1.5436853
#>  [66,] 5.4660967 2.1390520
#>  [67,] 5.2264946 2.3786541
#>  [68,] 5.4823432 2.1228055
#>  [69,] 6.2495358 1.3556129
#>  [70,] 6.0099337 1.5952150
#>  [71,] 4.9926463 2.6125023
#>  [72,] 4.9469835 2.6581652
#>  [73,] 6.2828577 1.3222910
#>  [74,] 5.9867407 1.6184080
#>  [75,] 6.8449025 0.7602461
#>  [76,] 6.0304042 1.5747445
#>  [77,] 5.9679218 1.6372269
#>  [78,] 7.0662968 0.5388519
#>  [79,] 7.6051487 0.0000000
#>  [80,] 5.2262809 2.3788678
#>  [81,] 6.8217095 0.7834391
#>  [82,] 4.4391908 3.1659578
#>  [83,] 6.5842104 1.0209383
#>  [84,] 6.3060507 1.2990980
#>  [85,] 6.2435683 1.3615804
#>  [86,] 7.0038144 0.6013343
#>  [87,] 7.0662968 0.5388519
#>  [88,] 6.2203753 1.3847734
#>  [89,] 5.8007712 1.8043774
#>  [90,] 7.0038144 0.6013343
#>  [91,] 7.0662968 0.5388519
#>  [92,] 6.0453585 1.5597902
#>  [93,] 6.2203753 1.3847734
#>  [94,] 6.3242585 1.2808902
#>  [95,] 6.2495358 1.3556129
#>  [96,] 6.0777100 1.5274387
#>  [97,] 7.6051487 0.0000000
#>  [98,] 4.9212773 2.6838714
#>  [99,] 3.9202365 3.6849122
#> [100,] 7.0662968 0.5388519
#> [101,] 6.0453585 1.5597902
#> [102,] 6.3060507 1.2990980
#> [103,] 5.8239643 1.7811844
#> [104,] 6.2621735 1.3429752
#> [105,] 6.0453585 1.5597902
#> [106,] 6.1032616 1.5018870
#> [107,] 5.8239643 1.7811844
#> [108,] 5.4710818 2.1340669
#> [109,] 5.4763757 2.1287730
#> [110,] 5.6299937 1.9751550
#> [111,] 6.2495358 1.3556129
#> [112,] 6.3428638 1.2622849
#> [113,] 6.0029873 1.6021614
#> [114,] 5.4272047 2.1779440
#> [115,] 6.3060507 1.2990980
#> [116,] 5.7233216 1.8818270
#> [117,] 6.0304042 1.5747445
#> [118,] 5.2450998 2.3600489
#> [119,] 5.8425695 1.7625792
#> [120,] 6.8449025 0.7602461
#> [121,] 6.8449025 0.7602461
#> [122,] 6.8449025 0.7602461
#> [123,] 6.8449025 0.7602461
#> [124,] 5.3037176 2.3014311
#> [125,] 5.2103897 2.3947590
#> [126,] 6.7701799 0.8349688
#> [127,] 6.2313280 1.3738207
#> [128,] 6.2495358 1.3556129
#> [129,] 5.7106839 1.8944648
#> [130,] 5.4710818 2.1340669
#> [131,] 6.2495358 1.3556129
#> [132,] 6.8449025 0.7602461
#> [133,] 7.6051487 0.0000000
#> [134,] 5.3033202 2.3018285
#> [135,] 6.8449025 0.7602461
#> [136,] 6.8449025 0.7602461
#> [137,] 5.7106839 1.8944648
#> [138,] 6.8635078 0.7416409
#> [139,] 4.2678250 3.3373237
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.16842408 0.83157592
#>   [2,] 0.16933747 0.83066253
#>   [3,] 0.40354517 0.59645483
#>   [4,] 0.16933747 0.83066253
#>   [5,] 0.25054309 0.74945691
#>   [6,] 0.42227470 0.57772530
#>   [7,] 0.26833649 0.73166351
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.20084272 0.79915728
#>  [10,] 0.34010174 0.65989826
#>  [11,] 0.33891184 0.66108816
#>  [12,] 0.00000000 1.00000000
#>  [13,] 0.27912741 0.72087259
#>  [14,] 0.44072361 0.55927639
#>  [15,] 0.20084272 0.79915728
#>  [16,] 0.23176131 0.76823869
#>  [17,] 0.25054309 0.74945691
#>  [18,] 0.14992293 0.85007707
#>  [19,] 0.07085356 0.92914644
#>  [20,] 0.24504704 0.75495296
#>  [21,] 0.17994741 0.82005259
#>  [22,] 0.07085356 0.92914644
#>  [23,] 0.09996467 0.90003533
#>  [24,] 0.07085356 0.92914644
#>  [25,] 0.16842408 0.83157592
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.27169628 0.72830372
#>  [28,] 0.40545926 0.59454074
#>  [29,] 0.10301431 0.89698569
#>  [30,] 0.20270246 0.79729754
#>  [31,] 0.27746568 0.72253432
#>  [32,] 0.20706295 0.79293705
#>  [33,] 0.18208367 0.81791633
#>  [34,] 0.23617405 0.76382595
#>  [35,] 0.00000000 1.00000000
#>  [36,] 0.09996467 0.90003533
#>  [37,] 0.23617405 0.76382595
#>  [38,] 0.19748293 0.80251707
#>  [39,] 0.13424304 0.86575696
#>  [40,] 0.00000000 1.00000000
#>  [41,] 0.09996467 0.90003533
#>  [42,] 0.09996467 0.90003533
#>  [43,] 0.09996467 0.90003533
#>  [44,] 0.09996467 0.90003533
#>  [45,] 0.09996467 0.90003533
#>  [46,] 0.07085356 0.92914644
#>  [47,] 0.09996467 0.90003533
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.29187362 0.70812638
#>  [50,] 0.20053257 0.79946743
#>  [51,] 0.10301431 0.89698569
#>  [52,] 0.25731872 0.74268128
#>  [53,] 0.35728339 0.64271661
#>  [54,] 0.28373816 0.71626184
#>  [55,] 0.20270246 0.79729754
#>  [56,] 0.27552236 0.72447764
#>  [57,] 0.27552236 0.72447764
#>  [58,] 0.20466880 0.79533120
#>  [59,] 0.37231644 0.62768356
#>  [60,] 0.30554685 0.69445315
#>  [61,] 0.17147373 0.82852627
#>  [62,] 0.23512109 0.76487891
#>  [63,] 0.10087805 0.89912195
#>  [64,] 1.00000000 0.00000000
#>  [65,] 0.79702102 0.20297898
#>  [66,] 0.71873633 0.28126367
#>  [67,] 0.68723108 0.31276892
#>  [68,] 0.72087259 0.27912741
#>  [69,] 0.82175064 0.17824936
#>  [70,] 0.79024539 0.20975461
#>  [71,] 0.65648241 0.34351759
#>  [72,] 0.65047821 0.34952179
#>  [73,] 0.82613213 0.17386787
#>  [74,] 0.78719575 0.21280425
#>  [75,] 0.90003533 0.09996467
#>  [76,] 0.79293705 0.20706295
#>  [77,] 0.78472125 0.21527875
#>  [78,] 0.92914644 0.07085356
#>  [79,] 1.00000000 0.00000000
#>  [80,] 0.68720299 0.31279701
#>  [81,] 0.89698569 0.10301431
#>  [82,] 0.58370862 0.41629138
#>  [83,] 0.86575696 0.13424304
#>  [84,] 0.82918177 0.17081823
#>  [85,] 0.82096597 0.17903403
#>  [86,] 0.92093064 0.07906936
#>  [87,] 0.92914644 0.07085356
#>  [88,] 0.81791633 0.18208367
#>  [89,] 0.76274265 0.23725735
#>  [90,] 0.92093064 0.07906936
#>  [91,] 0.92914644 0.07085356
#>  [92,] 0.79490340 0.20509660
#>  [93,] 0.81791633 0.18208367
#>  [94,] 0.83157592 0.16842408
#>  [95,] 0.82175064 0.17824936
#>  [96,] 0.79915728 0.20084272
#>  [97,] 1.00000000 0.00000000
#>  [98,] 0.64709810 0.35290190
#>  [99,] 0.51547138 0.48452862
#> [100,] 0.92914644 0.07085356
#> [101,] 0.79490340 0.20509660
#> [102,] 0.82918177 0.17081823
#> [103,] 0.76579229 0.23420771
#> [104,] 0.82341237 0.17658763
#> [105,] 0.79490340 0.20509660
#> [106,] 0.80251707 0.19748293
#> [107,] 0.76579229 0.23420771
#> [108,] 0.71939183 0.28060817
#> [109,] 0.72008792 0.27991208
#> [110,] 0.74028713 0.25971287
#> [111,] 0.82175064 0.17824936
#> [112,] 0.83402232 0.16597768
#> [113,] 0.78933201 0.21066799
#> [114,] 0.71362243 0.28637757
#> [115,] 0.82918177 0.17081823
#> [116,] 0.75255881 0.24744119
#> [117,] 0.79293705 0.20706295
#> [118,] 0.68967748 0.31032252
#> [119,] 0.76823869 0.23176131
#> [120,] 0.90003533 0.09996467
#> [121,] 0.90003533 0.09996467
#> [122,] 0.90003533 0.09996467
#> [123,] 0.90003533 0.09996467
#> [124,] 0.69738513 0.30261487
#> [125,] 0.68511345 0.31488655
#> [126,] 0.89021006 0.10978994
#> [127,] 0.81935649 0.18064351
#> [128,] 0.82175064 0.17824936
#> [129,] 0.75089708 0.24910292
#> [130,] 0.71939183 0.28060817
#> [131,] 0.82175064 0.17824936
#> [132,] 0.90003533 0.09996467
#> [133,] 1.00000000 0.00000000
#> [134,] 0.69733288 0.30266712
#> [135,] 0.90003533 0.09996467
#> [136,] 0.90003533 0.09996467
#> [137,] 0.75089708 0.24910292
#> [138,] 0.90248174 0.09751826
#> [139,] 0.56117575 0.43882425
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 3.0515792 7.1752890 6.3857613 0.0000000 1.6082976 0.0000000 2.0791748 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 2.0647945 1.0421079 0.0000000 0.0000000 1.7312378 2.4459052 0.0000000 4.2731257 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 2.1210648 0.6551813 0.0000000 0.0000000 2.2890612 0.0000000 0.8426788 2.3989963 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 0.0000000 2.2154331 0.0000000 0.5842672 3.7435763 6.6863936 0.0000000 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 2.0566526 4.6680527 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 7.7749538 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 1.8245373 3.6775213 0.0000000 0.0000000 0.0000000 6.2181280 5.1386652 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 0.0000000 2.7513229 0.0000000 0.0000000 2.6565853 0.0000000 1.8355962 
#>       V60        V7        V8        V9 
#> 0.8057001 0.0000000 0.0000000 7.1983591 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 76 63 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V45        V9       V11       V36       V12       V51       V52        V4 
#> 7.7749538 7.1983591 7.1752890 6.6863936 6.3857613 6.2181280 5.1386652 4.6680527 
#>       V23       V35       V48       V10       V55       V58       V21       V30 
#> 4.2731257 3.7435763 3.6775213 3.0515792 2.7513229 2.6565853 2.4459052 2.3989963 
#>       V28       V32       V24       V16       V17       V39        V6       V47 
#> 2.2890612 2.2154331 2.1210648 2.0791748 2.0647945 2.0566526 1.8355962 1.8245373 
#>       V20       V14       V18        V3       V60       V25       V34        V1 
#> 1.7312378 1.6082976 1.0421079 0.8426788 0.8057001 0.6551813 0.5842672 0.0000000 
#>       V13       V15       V19        V2       V22       V26       V27       V29 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V31       V33       V37       V38       V40       V41       V42       V43 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V44       V46       V49        V5       V50       V53       V54       V56 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V57       V59        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058