Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V12>=0.1677 87 31 M (0.64367816 0.35632184)  
#>      4) V51>=0.0136 47  5 M (0.89361702 0.10638298)  
#>        8) V9>=0.10915 40  1 M (0.97500000 0.02500000) *
#>        9) V9< 0.10915 7  3 R (0.42857143 0.57142857) *
#>      5) V51< 0.0136 40 14 R (0.35000000 0.65000000)  
#>       10) V23>=0.7906 10  1 M (0.90000000 0.10000000) *
#>       11) V23< 0.7906 30  5 R (0.16666667 0.83333333) *
#>    3) V12< 0.1677 52  6 R (0.11538462 0.88461538)  
#>      6) V48>=0.1309 8  3 M (0.62500000 0.37500000) *
#>      7) V48< 0.1309 44  1 R (0.02272727 0.97727273) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V11>=0.17975 95 25 M (0.73684211 0.26315789)  
#>      4) V36< 0.564 73  8 M (0.89041096 0.10958904)  
#>        8) V47>=0.06235 64  2 M (0.96875000 0.03125000) *
#>        9) V47< 0.06235 9  3 R (0.33333333 0.66666667) *
#>      5) V36>=0.564 22  5 R (0.22727273 0.77272727)  
#>       10) V17< 0.3251 7  2 M (0.71428571 0.28571429) *
#>       11) V17>=0.3251 15  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.17975 44  7 R (0.15909091 0.84090909)  
#>      6) V4>=0.0561 7  2 M (0.71428571 0.28571429) *
#>      7) V4< 0.0561 37  2 R (0.05405405 0.94594595) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V36< 0.3917 81 24 M (0.70370370 0.29629630)  
#>      4) V11>=0.1168 69 13 M (0.81159420 0.18840580)  
#>        8) V16< 0.76775 61  6 M (0.90163934 0.09836066)  
#>         16) V32>=0.24785 54  2 M (0.96296296 0.03703704) *
#>         17) V32< 0.24785 7  3 R (0.42857143 0.57142857) *
#>        9) V16>=0.76775 8  1 R (0.12500000 0.87500000) *
#>      5) V11< 0.1168 12  1 R (0.08333333 0.91666667) *
#>    3) V36>=0.3917 58 16 R (0.27586207 0.72413793)  
#>      6) V45>=0.35835 10  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.35835 48  6 R (0.12500000 0.87500000)  
#>       14) V20>=0.78785 8  3 M (0.62500000 0.37500000) *
#>       15) V20< 0.78785 40  1 R (0.02500000 0.97500000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V49>=0.04705 81 21 M (0.74074074 0.25925926)  
#>      4) V17< 0.51 55  5 M (0.90909091 0.09090909)  
#>        8) V51>=0.00695 48  1 M (0.97916667 0.02083333) *
#>        9) V51< 0.00695 7  3 R (0.42857143 0.57142857) *
#>      5) V17>=0.51 26 10 R (0.38461538 0.61538462)  
#>       10) V36< 0.19655 9  0 M (1.00000000 0.00000000) *
#>       11) V36>=0.19655 17  1 R (0.05882353 0.94117647) *
#>    3) V49< 0.04705 58 16 R (0.27586207 0.72413793)  
#>      6) V54>=0.02155 7  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.02155 51  9 R (0.17647059 0.82352941)  
#>       14) V2>=0.06875 7  1 M (0.85714286 0.14285714) *
#>       15) V2< 0.06875 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V36< 0.4781 98 28 M (0.71428571 0.28571429)  
#>     4) V11>=0.1119 87 17 M (0.80459770 0.19540230)  
#>       8) V17< 0.9445 78  9 M (0.88461538 0.11538462) *
#>       9) V17>=0.9445 9  1 R (0.11111111 0.88888889) *
#>     5) V11< 0.1119 11  0 R (0.00000000 1.00000000) *
#>   3) V36>=0.4781 41  7 R (0.17073171 0.82926829)  
#>     6) V37>=0.6663 8  2 M (0.75000000 0.25000000) *
#>     7) V37< 0.6663 33  1 R (0.03030303 0.96969697) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V36< 0.52905 104 31 M (0.70192308 0.29807692)  
#>      4) V24>=0.611 64  6 M (0.90625000 0.09375000)  
#>        8) V29>=0.416 57  2 M (0.96491228 0.03508772) *
#>        9) V29< 0.416 7  3 R (0.42857143 0.57142857) *
#>      5) V24< 0.611 40 15 R (0.37500000 0.62500000)  
#>       10) V26< 0.2282 10  0 M (1.00000000 0.00000000) *
#>       11) V26>=0.2282 30  5 R (0.16666667 0.83333333)  
#>         22) V42>=0.27455 7  2 M (0.71428571 0.28571429) *
#>         23) V42< 0.27455 23  0 R (0.00000000 1.00000000) *
#>    3) V36>=0.52905 35  3 R (0.08571429 0.91428571) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V48>=0.04855 100 24 M (0.76000000 0.24000000)  
#>      4) V8>=0.1017 62  4 M (0.93548387 0.06451613) *
#>      5) V8< 0.1017 38 18 R (0.47368421 0.52631579)  
#>       10) V31< 0.3609 12  0 M (1.00000000 0.00000000) *
#>       11) V31>=0.3609 26  6 R (0.23076923 0.76923077)  
#>         22) V18< 0.28735 8  2 M (0.75000000 0.25000000) *
#>         23) V18>=0.28735 18  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.04855 39  3 R (0.07692308 0.92307692) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V10>=0.14975 84 18 M (0.78571429 0.21428571)  
#>      4) V36< 0.564 62  7 M (0.88709677 0.11290323) *
#>      5) V36>=0.564 22 11 M (0.50000000 0.50000000)  
#>       10) V20< 0.2225 11  0 M (1.00000000 0.00000000) *
#>       11) V20>=0.2225 11  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.14975 55 12 R (0.21818182 0.78181818)  
#>      6) V29< 0.4632 12  2 M (0.83333333 0.16666667) *
#>      7) V29>=0.4632 43  2 R (0.04651163 0.95348837) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V52>=0.0047 116 41 M (0.64655172 0.35344828)  
#>      4) V32>=0.0825 104 29 M (0.72115385 0.27884615)  
#>        8) V12>=0.08385 96 22 M (0.77083333 0.22916667)  
#>         16) V23>=0.74575 43  1 M (0.97674419 0.02325581) *
#>         17) V23< 0.74575 53 21 M (0.60377358 0.39622642)  
#>           34) V18< 0.3189 32  5 M (0.84375000 0.15625000)  
#>             68) V51>=0.0087 25  0 M (1.00000000 0.00000000) *
#>             69) V51< 0.0087 7  2 R (0.28571429 0.71428571) *
#>           35) V18>=0.3189 21  5 R (0.23809524 0.76190476)  
#>             70) V12>=0.3847 7  2 M (0.71428571 0.28571429) *
#>             71) V12< 0.3847 14  0 R (0.00000000 1.00000000) *
#>        9) V12< 0.08385 8  1 R (0.12500000 0.87500000) *
#>      5) V32< 0.0825 12  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0047 23  1 R (0.04347826 0.95652174) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V4>=0.052 69 17 M (0.75362319 0.24637681)  
#>      4) V50>=0.01825 39  0 M (1.00000000 0.00000000) *
#>      5) V50< 0.01825 30 13 R (0.43333333 0.56666667)  
#>       10) V3< 0.05765 11  0 M (1.00000000 0.00000000) *
#>       11) V3>=0.05765 19  2 R (0.10526316 0.89473684) *
#>    3) V4< 0.052 70 22 R (0.31428571 0.68571429)  
#>      6) V20>=0.51955 25  9 M (0.64000000 0.36000000)  
#>       12) V30>=0.5136 15  1 M (0.93333333 0.06666667) *
#>       13) V30< 0.5136 10  2 R (0.20000000 0.80000000) *
#>      7) V20< 0.51955 45  6 R (0.13333333 0.86666667)  
#>       14) V32>=0.54695 11  5 M (0.54545455 0.45454545) *
#>       15) V32< 0.54695 34  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7586613 0.9301451 0.9035812 0.7607245 0.6931697 0.7167864 0.5806716
#>  [8] 0.7984953 0.8451537 0.6731706
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.0325027 5.6280567
#>   [2,] 2.2243206 5.4362388
#>   [3,] 2.2503263 5.4102331
#>   [4,] 1.3791669 6.2813925
#>   [5,] 2.3093119 5.3512475
#>   [6,] 2.4218101 5.2387493
#>   [7,] 1.2738414 6.3867181
#>   [8,] 1.2538422 6.4067172
#>   [9,] 1.6058782 6.0546812
#>  [10,] 2.2243206 5.4362388
#>  [11,] 1.4258253 6.2347341
#>  [12,] 2.2084514 5.4521080
#>  [13,] 2.0325027 5.6280567
#>  [14,] 3.1146317 4.5459277
#>  [15,] 0.6731706 6.9873888
#>  [16,] 2.2503263 5.4102331
#>  [17,] 2.5073962 5.1531632
#>  [18,] 2.7921656 4.8683938
#>  [19,] 1.4338951 6.2266643
#>  [20,] 0.6731706 6.9873888
#>  [21,] 1.6469315 6.0136279
#>  [22,] 2.3135374 5.3470220
#>  [23,] 2.1648356 5.4957238
#>  [24,] 1.6033157 6.0572437
#>  [25,] 2.3374763 5.3230831
#>  [26,] 2.5268960 5.1336634
#>  [27,] 0.6731706 6.9873888
#>  [28,] 1.3899570 6.2706024
#>  [29,] 0.6931697 6.9673897
#>  [30,] 1.4258253 6.2347341
#>  [31,] 0.6731706 6.9873888
#>  [32,] 1.3663403 6.2942191
#>  [33,] 1.3663403 6.2942191
#>  [34,] 2.4219055 5.2386539
#>  [35,] 2.8109988 4.8495606
#>  [36,] 1.5152817 6.1452777
#>  [37,] 1.2974580 6.3631014
#>  [38,] 2.3135374 5.3470220
#>  [39,] 2.3135374 5.3470220
#>  [40,] 0.6731706 6.9873888
#>  [41,] 0.6731706 6.9873888
#>  [42,] 0.7167864 6.9437730
#>  [43,] 2.7077364 4.9528230
#>  [44,] 1.2738414 6.3867181
#>  [45,] 0.6931697 6.9673897
#>  [46,] 0.7167864 6.9437730
#>  [47,] 0.0000000 7.6605594
#>  [48,] 1.5152817 6.1452777
#>  [49,] 1.5152817 6.1452777
#>  [50,] 2.0523375 5.6082219
#>  [51,] 1.3899570 6.2706024
#>  [52,] 2.9824825 4.6780769
#>  [53,] 3.1625354 4.4980240
#>  [54,] 1.7286403 5.9319191
#>  [55,] 2.0325027 5.6280567
#>  [56,] 2.8776564 4.7829030
#>  [57,] 1.4258253 6.2347341
#>  [58,] 1.3393329 6.3212265
#>  [59,] 0.6731706 6.9873888
#>  [60,] 1.5592197 6.1013397
#>  [61,] 2.3604354 5.3001240
#>  [62,] 2.3135374 5.3470220
#>  [63,] 1.5619401 6.0986193
#>  [64,] 0.7167864 6.9437730
#>  [65,] 5.4595202 2.2010392
#>  [66,] 6.0567444 1.6038150
#>  [67,] 6.8998349 0.7607245
#>  [68,] 4.7179108 2.9426486
#>  [69,] 5.5605020 2.1000574
#>  [70,] 5.3890182 2.2715412
#>  [71,] 5.4480039 2.2125555
#>  [72,] 5.5098779 2.1506815
#>  [73,] 5.1998216 2.4607378
#>  [74,] 4.9967610 2.6637984
#>  [75,] 4.5792335 3.0813259
#>  [76,] 6.0838076 1.5767518
#>  [77,] 7.0798878 0.5806716
#>  [78,] 6.8620641 0.7984953
#>  [79,] 6.0546812 1.6058782
#>  [80,] 7.6605594 0.0000000
#>  [81,] 7.6605594 0.0000000
#>  [82,] 6.4067172 1.2538422
#>  [83,] 6.7569782 0.9035812
#>  [84,] 6.0838076 1.5767518
#>  [85,] 5.9962537 1.6643057
#>  [86,] 6.7569782 0.9035812
#>  [87,] 5.1336634 2.5268960
#>  [88,] 6.1452777 1.5152817
#>  [89,] 6.1452777 1.5152817
#>  [90,] 5.3001240 2.3604354
#>  [91,] 5.1998216 2.4607378
#>  [92,] 6.1888936 1.4716659
#>  [93,] 6.7304143 0.9301451
#>  [94,] 5.4362388 2.2243206
#>  [95,] 7.6605594 0.0000000
#>  [96,] 6.8620641 0.7984953
#>  [97,] 6.3631014 1.2974580
#>  [98,] 6.7569782 0.9035812
#>  [99,] 5.9118245 1.7487349
#> [100,] 4.7512166 2.9093428
#> [101,] 5.1265993 2.5339601
#> [102,] 4.7153483 2.9452111
#> [103,] 6.9437730 0.7167864
#> [104,] 5.4919419 2.1686175
#> [105,] 7.0798878 0.5806716
#> [106,] 6.8998349 0.7607245
#> [107,] 6.0401918 1.6203676
#> [108,] 6.7304143 0.9301451
#> [109,] 5.2151327 2.4454267
#> [110,] 6.2066652 1.4538942
#> [111,] 6.1422351 1.5183243
#> [112,] 6.2266643 1.4338951
#> [113,] 4.7597857 2.9007737
#> [114,] 5.4155821 2.2449773
#> [115,] 5.2110285 2.4495309
#> [116,] 6.9018981 0.7586613
#> [117,] 7.6605594 0.0000000
#> [118,] 7.6605594 0.0000000
#> [119,] 7.6605594 0.0000000
#> [120,] 7.6605594 0.0000000
#> [121,] 7.6605594 0.0000000
#> [122,] 7.6605594 0.0000000
#> [123,] 6.9873888 0.6731706
#> [124,] 6.9873888 0.6731706
#> [125,] 6.2266643 1.4338951
#> [126,] 7.6605594 0.0000000
#> [127,] 6.2287275 1.4318319
#> [128,] 7.6605594 0.0000000
#> [129,] 7.6605594 0.0000000
#> [130,] 7.0798878 0.5806716
#> [131,] 7.6605594 0.0000000
#> [132,] 7.6605594 0.0000000
#> [133,] 6.7569782 0.9035812
#> [134,] 7.6605594 0.0000000
#> [135,] 7.6605594 0.0000000
#> [136,] 6.8998349 0.7607245
#> [137,] 7.6605594 0.0000000
#> [138,] 5.2960199 2.3645395
#> [139,] 5.6459927 2.0145667
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.26532040 0.73467960
#>   [2,] 0.29036007 0.70963993
#>   [3,] 0.29375483 0.70624517
#>   [4,] 0.18003475 0.81996525
#>   [5,] 0.30145474 0.69854526
#>   [6,] 0.31614011 0.68385989
#>   [7,] 0.16628568 0.83371432
#>   [8,] 0.16367502 0.83632498
#>   [9,] 0.20962936 0.79037064
#>  [10,] 0.29036007 0.70963993
#>  [11,] 0.18612548 0.81387452
#>  [12,] 0.28828853 0.71171147
#>  [13,] 0.26532040 0.73467960
#>  [14,] 0.40658019 0.59341981
#>  [15,] 0.08787486 0.91212514
#>  [16,] 0.29375483 0.70624517
#>  [17,] 0.32731242 0.67268758
#>  [18,] 0.36448587 0.63551413
#>  [19,] 0.18717890 0.81282110
#>  [20,] 0.08787486 0.91212514
#>  [21,] 0.21498841 0.78501159
#>  [22,] 0.30200632 0.69799368
#>  [23,] 0.28259498 0.71740502
#>  [24,] 0.20929486 0.79070514
#>  [25,] 0.30513128 0.69486872
#>  [26,] 0.32985790 0.67014210
#>  [27,] 0.08787486 0.91212514
#>  [28,] 0.18144328 0.81855672
#>  [29,] 0.09048553 0.90951447
#>  [30,] 0.18612548 0.81387452
#>  [31,] 0.08787486 0.91212514
#>  [32,] 0.17836039 0.82163961
#>  [33,] 0.17836039 0.82163961
#>  [34,] 0.31615257 0.68384743
#>  [35,] 0.36694432 0.63305568
#>  [36,] 0.19780300 0.80219700
#>  [37,] 0.16936857 0.83063143
#>  [38,] 0.30200632 0.69799368
#>  [39,] 0.30200632 0.69799368
#>  [40,] 0.08787486 0.91212514
#>  [41,] 0.08787486 0.91212514
#>  [42,] 0.09356842 0.90643158
#>  [43,] 0.35346458 0.64653542
#>  [44,] 0.16628568 0.83371432
#>  [45,] 0.09048553 0.90951447
#>  [46,] 0.09356842 0.90643158
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.19780300 0.80219700
#>  [49,] 0.19780300 0.80219700
#>  [50,] 0.26790961 0.73209039
#>  [51,] 0.18144328 0.81855672
#>  [52,] 0.38932960 0.61067040
#>  [53,] 0.41283348 0.58716652
#>  [54,] 0.22565458 0.77434542
#>  [55,] 0.26532040 0.73467960
#>  [56,] 0.37564572 0.62435428
#>  [57,] 0.18612548 0.81387452
#>  [58,] 0.17483487 0.82516513
#>  [59,] 0.08787486 0.91212514
#>  [60,] 0.20353863 0.79646137
#>  [61,] 0.30812833 0.69187167
#>  [62,] 0.30200632 0.69799368
#>  [63,] 0.20389374 0.79610626
#>  [64,] 0.09356842 0.90643158
#>  [65,] 0.71267905 0.28732095
#>  [66,] 0.79063996 0.20936004
#>  [67,] 0.90069596 0.09930404
#>  [68,] 0.61587027 0.38412973
#>  [69,] 0.72586109 0.27413891
#>  [70,] 0.70347581 0.29652419
#>  [71,] 0.71117572 0.28882428
#>  [72,] 0.71925268 0.28074732
#>  [73,] 0.67877832 0.32122168
#>  [74,] 0.65227104 0.34772896
#>  [75,] 0.59776751 0.40223249
#>  [76,] 0.79417276 0.20582724
#>  [77,] 0.92419984 0.07580016
#>  [78,] 0.89576541 0.10423459
#>  [79,] 0.79037064 0.20962936
#>  [80,] 1.00000000 0.00000000
#>  [81,] 1.00000000 0.00000000
#>  [82,] 0.83632498 0.16367502
#>  [83,] 0.88204762 0.11795238
#>  [84,] 0.79417276 0.20582724
#>  [85,] 0.78274358 0.21725642
#>  [86,] 0.88204762 0.11795238
#>  [87,] 0.67014210 0.32985790
#>  [88,] 0.80219700 0.19780300
#>  [89,] 0.80219700 0.19780300
#>  [90,] 0.69187167 0.30812833
#>  [91,] 0.67877832 0.32122168
#>  [92,] 0.80789055 0.19210945
#>  [93,] 0.87858001 0.12141999
#>  [94,] 0.70963993 0.29036007
#>  [95,] 1.00000000 0.00000000
#>  [96,] 0.89576541 0.10423459
#>  [97,] 0.83063143 0.16936857
#>  [98,] 0.88204762 0.11795238
#>  [99,] 0.77172230 0.22827770
#> [100,] 0.62021797 0.37978203
#> [101,] 0.66921997 0.33078003
#> [102,] 0.61553577 0.38446423
#> [103,] 0.90643158 0.09356842
#> [104,] 0.71691134 0.28308866
#> [105,] 0.92419984 0.07580016
#> [106,] 0.90069596 0.09930404
#> [107,] 0.78847920 0.21152080
#> [108,] 0.87858001 0.12141999
#> [109,] 0.68077700 0.31922300
#> [110,] 0.81021044 0.18978956
#> [111,] 0.80179982 0.19820018
#> [112,] 0.81282110 0.18717890
#> [113,] 0.62133657 0.37866343
#> [114,] 0.70694342 0.29305658
#> [115,] 0.68024125 0.31975875
#> [116,] 0.90096529 0.09903471
#> [117,] 1.00000000 0.00000000
#> [118,] 1.00000000 0.00000000
#> [119,] 1.00000000 0.00000000
#> [120,] 1.00000000 0.00000000
#> [121,] 1.00000000 0.00000000
#> [122,] 1.00000000 0.00000000
#> [123,] 0.91212514 0.08787486
#> [124,] 0.91212514 0.08787486
#> [125,] 0.81282110 0.18717890
#> [126,] 1.00000000 0.00000000
#> [127,] 0.81309042 0.18690958
#> [128,] 1.00000000 0.00000000
#> [129,] 1.00000000 0.00000000
#> [130,] 0.92419984 0.07580016
#> [131,] 1.00000000 0.00000000
#> [132,] 1.00000000 0.00000000
#> [133,] 0.88204762 0.11795238
#> [134,] 1.00000000 0.00000000
#> [135,] 1.00000000 0.00000000
#> [136,] 0.90069596 0.09930404
#> [137,] 1.00000000 0.00000000
#> [138,] 0.69133592 0.30866408
#> [139,] 0.73702094 0.26297906
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  4.2964308  9.3572910  5.7833937  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  1.9373299  4.6770793  2.8848532  0.0000000  1.4372950  4.6931395  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  2.9407032  2.5027712  0.0000000  1.8763568  0.0000000  0.0000000 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  2.9770693  1.8869817  1.0917071  1.4178875  4.0169888  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000 15.8123352  1.1619459  0.0000000  0.0000000  3.4687894  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.9864275  0.0000000  0.0000000  2.8775536  0.0000000  1.4893419 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  4.7570001  2.7927479  0.0000000  1.8422311  4.3283129  2.9653369  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  1.5960729  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  1.4663579  0.6782687 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 75 64 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V36        V11        V12        V48        V20        V17        V51 
#> 15.8123352  9.3572910  5.7833937  4.7570001  4.6931395  4.6770793  4.3283129 
#>        V10        V32         V4        V29        V52        V23        V18 
#>  4.2964308  4.0169888  3.4687894  2.9770693  2.9653369  2.9407032  2.8848532 
#>        V45        V49        V24        V16         V3        V26        V50 
#>  2.8775536  2.7927479  2.5027712  1.9373299  1.8869817  1.8763568  1.8422311 
#>        V54        V47         V8         V2        V31        V37        V30 
#>  1.5960729  1.4893419  1.4663579  1.4372950  1.4178875  1.1619459  1.0917071 
#>        V42         V9         V1        V13        V14        V15        V19 
#>  0.9864275  0.6782687  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V21        V22        V25        V27        V28        V33        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V38        V39        V40        V41        V43        V44 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V46         V5        V53        V55        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V59         V6        V60         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913