Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 65 M (0.53237410 0.46762590)  
#>   2) V10>=0.1663 84 19 M (0.77380952 0.22619048)  
#>     4) V12>=0.1271 71  9 M (0.87323944 0.12676056) *
#>     5) V12< 0.1271 13  3 R (0.23076923 0.76923077) *
#>   3) V10< 0.1663 55  9 R (0.16363636 0.83636364)  
#>     6) V46>=0.2079 9  3 M (0.66666667 0.33333333) *
#>     7) V46< 0.2079 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V49>=0.0755 37  4 M (0.89189189 0.10810811)  
#>      4) V39< 0.7198 30  0 M (1.00000000 0.00000000) *
#>      5) V39>=0.7198 7  3 R (0.42857143 0.57142857) *
#>    3) V49< 0.0755 102 29 R (0.28431373 0.71568627)  
#>      6) V23>=0.7109 41 19 M (0.53658537 0.46341463)  
#>       12) V13>=0.18135 27  6 M (0.77777778 0.22222222)  
#>         24) V11>=0.1587 17  0 M (1.00000000 0.00000000) *
#>         25) V11< 0.1587 10  4 R (0.40000000 0.60000000) *
#>       13) V13< 0.18135 14  1 R (0.07142857 0.92857143) *
#>      7) V23< 0.7109 61  7 R (0.11475410 0.88524590)  
#>       14) V15>=0.65775 7  3 M (0.57142857 0.42857143) *
#>       15) V15< 0.65775 54  3 R (0.05555556 0.94444444) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V11>=0.16095 98 40 M (0.59183673 0.40816327)  
#>      4) V18< 0.5362 49  7 M (0.85714286 0.14285714) *
#>      5) V18>=0.5362 49 16 R (0.32653061 0.67346939)  
#>       10) V37< 0.4705 25  9 M (0.64000000 0.36000000)  
#>         20) V17< 0.9452 17  1 M (0.94117647 0.05882353) *
#>         21) V17>=0.9452 8  0 R (0.00000000 1.00000000) *
#>       11) V37>=0.4705 24  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.16095 41  7 R (0.17073171 0.82926829)  
#>      6) V19>=0.82425 8  2 M (0.75000000 0.25000000) *
#>      7) V19< 0.82425 33  1 R (0.03030303 0.96969697) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 R (0.41726619 0.58273381)  
#>    2) V48>=0.07635 56 18 M (0.67857143 0.32142857)  
#>      4) V41< 0.66555 49 11 M (0.77551020 0.22448980)  
#>        8) V12>=0.1641 34  3 M (0.91176471 0.08823529) *
#>        9) V12< 0.1641 15  7 R (0.46666667 0.53333333) *
#>      5) V41>=0.66555 7  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.07635 83 20 R (0.24096386 0.75903614)  
#>      6) V28< 0.49705 20  5 M (0.75000000 0.25000000)  
#>       12) V36< 0.34105 12  0 M (1.00000000 0.00000000) *
#>       13) V36>=0.34105 8  3 R (0.37500000 0.62500000) *
#>      7) V28>=0.49705 63  5 R (0.07936508 0.92063492) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V55< 0.0067 63 12 M (0.80952381 0.19047619)  
#>      4) V48>=0.0331 56  6 M (0.89285714 0.10714286) *
#>      5) V48< 0.0331 7  1 R (0.14285714 0.85714286) *
#>    3) V55>=0.0067 76 24 R (0.31578947 0.68421053)  
#>      6) V49>=0.07225 18  1 M (0.94444444 0.05555556) *
#>      7) V49< 0.07225 58  7 R (0.12068966 0.87931034)  
#>       14) V23>=0.8234 9  4 M (0.55555556 0.44444444) *
#>       15) V23< 0.8234 49  2 R (0.04081633 0.95918367) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V31< 0.4119 58 13 M (0.77586207 0.22413793)  
#>      4) V33>=0.22535 43  1 M (0.97674419 0.02325581) *
#>      5) V33< 0.22535 15  3 R (0.20000000 0.80000000) *
#>    3) V31>=0.4119 81 24 R (0.29629630 0.70370370)  
#>      6) V11>=0.24225 23  7 M (0.69565217 0.30434783)  
#>       12) V17< 0.43475 16  0 M (1.00000000 0.00000000) *
#>       13) V17>=0.43475 7  0 R (0.00000000 1.00000000) *
#>      7) V11< 0.24225 58  8 R (0.13793103 0.86206897) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.4676259 0.5323741)  
#>    2) V28>=0.9055 22  0 M (1.0000000 0.0000000) *
#>    3) V28< 0.9055 117 43 R (0.3675214 0.6324786)  
#>      6) V52>=0.0115 56 20 M (0.6428571 0.3571429)  
#>       12) V36< 0.50915 37  5 M (0.8648649 0.1351351)  
#>         24) V1>=0.02085 30  0 M (1.0000000 0.0000000) *
#>         25) V1< 0.02085 7  2 R (0.2857143 0.7142857) *
#>       13) V36>=0.50915 19  4 R (0.2105263 0.7894737) *
#>      7) V52< 0.0115 61  7 R (0.1147541 0.8852459) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V9>=0.11025 96 20 M (0.79166667 0.20833333)  
#>      4) V36< 0.46485 65  5 M (0.92307692 0.07692308) *
#>      5) V36>=0.46485 31 15 M (0.51612903 0.48387097)  
#>       10) V45>=0.26515 15  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.26515 16  1 R (0.06250000 0.93750000) *
#>    3) V9< 0.11025 43  6 R (0.13953488 0.86046512)  
#>      6) V45>=0.2037 11  5 M (0.54545455 0.45454545) *
#>      7) V45< 0.2037 32  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V59>=0.01125 34  2 M (0.94117647 0.05882353) *
#>    3) V59< 0.01125 105 40 R (0.38095238 0.61904762)  
#>      6) V56< 0.00725 51 20 M (0.60784314 0.39215686)  
#>       12) V21>=0.5542 35  5 M (0.85714286 0.14285714) *
#>       13) V21< 0.5542 16  1 R (0.06250000 0.93750000) *
#>      7) V56>=0.00725 54  9 R (0.16666667 0.83333333)  
#>       14) V49>=0.0815 10  4 M (0.60000000 0.40000000) *
#>       15) V49< 0.0815 44  3 R (0.06818182 0.93181818) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.4532374 0.5467626)  
#>    2) V10>=0.1678 64 19 M (0.7031250 0.2968750)  
#>      4) V36< 0.48855 39  4 M (0.8974359 0.1025641) *
#>      5) V36>=0.48855 25 10 R (0.4000000 0.6000000)  
#>       10) V10>=0.4392 9  0 M (1.0000000 0.0000000) *
#>       11) V10< 0.4392 16  1 R (0.0625000 0.9375000) *
#>    3) V10< 0.1678 75 18 R (0.2400000 0.7600000)  
#>      6) V44>=0.16585 38 18 R (0.4736842 0.5263158)  
#>       12) V33< 0.50925 22  4 M (0.8181818 0.1818182)  
#>         24) V30< 0.63875 15  0 M (1.0000000 0.0000000) *
#>         25) V30>=0.63875 7  3 R (0.4285714 0.5714286) *
#>       13) V33>=0.50925 16  0 R (0.0000000 1.0000000) *
#>      7) V44< 0.16585 37  0 R (0.0000000 1.0000000) *
#> 
#> 
#> $weights
#>  [1] 0.6450753 0.6760528 0.8595990 0.6266012 0.6691535 0.7773397 0.7134534
#>  [8] 0.6395836 0.6634782 0.7937716
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.4224149 5.6416933
#>   [2,] 2.7544832 4.3096250
#>   [3,] 2.7918838 4.2722244
#>   [4,] 1.9112601 5.1528481
#>   [5,] 1.5287525 5.5353557
#>   [6,] 2.9380295 4.1260787
#>   [7,] 0.6691535 6.3949547
#>   [8,] 1.4169233 5.6471849
#>   [9,] 1.3326317 5.7314765
#>  [10,] 2.1681525 4.8959557
#>  [11,] 1.5230772 5.5410310
#>  [12,] 1.5046743 5.5594339
#>  [13,] 0.7773397 6.2867685
#>  [14,] 0.7134534 6.3506548
#>  [15,] 1.9846062 5.0795020
#>  [16,] 1.3142287 5.7498794
#>  [17,] 2.6172906 4.4468176
#>  [18,] 2.0221905 5.0419177
#>  [19,] 0.0000000 7.0641082
#>  [20,] 1.3087371 5.7553711
#>  [21,] 1.5356518 5.5284564
#>  [22,] 1.9351547 5.1289535
#>  [23,] 2.9380295 4.1260787
#>  [24,] 0.8595990 6.2045092
#>  [25,] 0.7773397 6.2867685
#>  [26,] 1.5287525 5.5353557
#>  [27,] 0.6266012 6.4375070
#>  [28,] 2.1683362 4.8957720
#>  [29,] 0.6691535 6.3949547
#>  [30,] 0.7134534 6.3506548
#>  [31,] 0.6691535 6.3949547
#>  [32,] 2.7659869 4.2981213
#>  [33,] 0.6395836 6.4245246
#>  [34,] 2.6172906 4.4468176
#>  [35,] 2.0984677 4.9656405
#>  [36,] 2.0858931 4.9782151
#>  [37,] 1.9538124 5.1102958
#>  [38,] 2.1991300 4.8649782
#>  [39,] 1.2846589 5.7794493
#>  [40,] 2.1763784 4.8877298
#>  [41,] 0.6691535 6.3949547
#>  [42,] 2.2126360 4.8514721
#>  [43,] 0.0000000 7.0641082
#>  [44,] 2.0460851 5.0180231
#>  [45,] 1.3769316 5.6871766
#>  [46,] 1.3326317 5.7314765
#>  [47,] 1.3769316 5.6871766
#>  [48,] 0.0000000 7.0641082
#>  [49,] 1.2846589 5.7794493
#>  [50,] 0.6691535 6.3949547
#>  [51,] 1.3087371 5.7553711
#>  [52,] 0.0000000 7.0641082
#>  [53,] 2.2402647 4.8238435
#>  [54,] 0.7134534 6.3506548
#>  [55,] 2.8318144 4.2322938
#>  [56,] 2.1683362 4.8957720
#>  [57,] 2.0221905 5.0419177
#>  [58,] 2.1099713 4.9541368
#>  [59,] 1.9592329 5.1048753
#>  [60,] 2.1991300 4.8649782
#>  [61,] 3.3810844 3.6830238
#>  [62,] 2.7050317 4.3590765
#>  [63,] 1.9902815 5.0738267
#>  [64,] 1.3452063 5.7189019
#>  [65,] 1.3326317 5.7314765
#>  [66,] 1.3395310 5.7245772
#>  [67,] 1.3530370 5.7110712
#>  [68,] 1.3030618 5.7610464
#>  [69,] 4.0577006 3.0064076
#>  [70,] 6.3949547 0.6691535
#>  [71,] 7.0641082 0.0000000
#>  [72,] 5.7189019 1.3452063
#>  [73,] 4.3899432 2.6741650
#>  [74,] 4.2567799 2.8073283
#>  [75,] 4.2083381 2.8557701
#>  [76,] 5.7314765 1.3326317
#>  [77,] 5.7240536 1.3400546
#>  [78,] 5.5733152 1.4907930
#>  [79,] 4.3516139 2.7124943
#>  [80,] 6.4006300 0.6634782
#>  [81,] 5.6232903 1.4408179
#>  [82,] 4.9966891 2.0674191
#>  [83,] 6.2867685 0.7773397
#>  [84,] 4.9467139 2.1173943
#>  [85,] 5.6232903 1.4408179
#>  [86,] 6.4375070 0.6266012
#>  [87,] 4.7580160 2.3060922
#>  [88,] 6.3506548 0.7134534
#>  [89,] 5.0111239 2.0529843
#>  [90,] 5.0549001 2.0092081
#>  [91,] 5.0789784 1.9851298
#>  [92,] 5.0054486 2.0586596
#>  [93,] 7.0641082 0.0000000
#>  [94,] 4.8534133 2.2106949
#>  [95,] 4.1344683 2.9296399
#>  [96,] 5.4271695 1.6369387
#>  [97,] 4.3111037 2.7530045
#>  [98,] 5.5568833 1.5072249
#>  [99,] 6.2703366 0.7937716
#> [100,] 5.6601673 1.4039409
#> [101,] 5.7614542 1.3026540
#> [102,] 5.5568833 1.5072249
#> [103,] 5.6746020 1.3895061
#> [104,] 5.0979760 1.9661322
#> [105,] 4.8972624 2.1668458
#> [106,] 4.4029256 2.6611826
#> [107,] 4.9656405 2.0984677
#> [108,] 5.6068584 1.4572498
#> [109,] 5.4929970 1.5711112
#> [110,] 6.2703366 0.7937716
#> [111,] 5.6232903 1.4408179
#> [112,] 6.3506548 0.7134534
#> [113,] 5.7240536 1.3400546
#> [114,] 5.7110712 1.3530370
#> [115,] 4.8005683 2.2635399
#> [116,] 4.8005683 2.2635399
#> [117,] 6.4375070 0.6266012
#> [118,] 4.3295067 2.7346015
#> [119,] 5.7240536 1.3400546
#> [120,] 5.7110712 1.3530370
#> [121,] 5.7614542 1.3026540
#> [122,] 6.3506548 0.7134534
#> [123,] 5.7240536 1.3400546
#> [124,] 5.7240536 1.3400546
#> [125,] 4.9472376 2.1168706
#> [126,] 6.3880554 0.6760528
#> [127,] 7.0641082 0.0000000
#> [128,] 7.0641082 0.0000000
#> [129,] 7.0641082 0.0000000
#> [130,] 5.6871766 1.3769316
#> [131,] 5.7245772 1.3395310
#> [132,] 7.0641082 0.0000000
#> [133,] 6.2867685 0.7773397
#> [134,] 6.2867685 0.7773397
#> [135,] 6.4375070 0.6266012
#> [136,] 7.0641082 0.0000000
#> [137,] 5.7610464 1.3030618
#> [138,] 6.4006300 0.6634782
#> [139,] 5.6746020 1.3895061
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.20135803 0.79864197
#>   [2,] 0.38992653 0.61007347
#>   [3,] 0.39522099 0.60477901
#>   [4,] 0.27055929 0.72944071
#>   [5,] 0.21641126 0.78358874
#>   [6,] 0.41590947 0.58409053
#>   [7,] 0.09472583 0.90527417
#>   [8,] 0.20058063 0.79941937
#>   [9,] 0.18864826 0.81135174
#>  [10,] 0.30692515 0.69307485
#>  [11,] 0.21560786 0.78439214
#>  [12,] 0.21300272 0.78699728
#>  [13,] 0.11004074 0.88995926
#>  [14,] 0.10099695 0.89900305
#>  [15,] 0.28094222 0.71905778
#>  [16,] 0.18604312 0.81395688
#>  [17,] 0.37050545 0.62949455
#>  [18,] 0.28626267 0.71373733
#>  [19,] 0.00000000 1.00000000
#>  [20,] 0.18526572 0.81473428
#>  [21,] 0.21738792 0.78261208
#>  [22,] 0.27394182 0.72605818
#>  [23,] 0.41590947 0.58409053
#>  [24,] 0.12168543 0.87831457
#>  [25,] 0.11004074 0.88995926
#>  [26,] 0.21641126 0.78358874
#>  [27,] 0.08870210 0.91129790
#>  [28,] 0.30695115 0.69304885
#>  [29,] 0.09472583 0.90527417
#>  [30,] 0.10099695 0.89900305
#>  [31,] 0.09472583 0.90527417
#>  [32,] 0.39155500 0.60844500
#>  [33,] 0.09053990 0.90946010
#>  [34,] 0.37050545 0.62949455
#>  [35,] 0.29706052 0.70293948
#>  [36,] 0.29528046 0.70471954
#>  [37,] 0.27658302 0.72341698
#>  [38,] 0.31131035 0.68868965
#>  [39,] 0.18185719 0.81814281
#>  [40,] 0.30808962 0.69191038
#>  [41,] 0.09472583 0.90527417
#>  [42,] 0.31322228 0.68677772
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.28964520 0.71035480
#>  [45,] 0.19491938 0.80508062
#>  [46,] 0.18864826 0.81135174
#>  [47,] 0.19491938 0.80508062
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.18185719 0.81814281
#>  [50,] 0.09472583 0.90527417
#>  [51,] 0.18526572 0.81473428
#>  [52,] 0.00000000 1.00000000
#>  [53,] 0.31713341 0.68286659
#>  [54,] 0.10099695 0.89900305
#>  [55,] 0.40087358 0.59912642
#>  [56,] 0.30695115 0.69304885
#>  [57,] 0.28626267 0.71373733
#>  [58,] 0.29868899 0.70131101
#>  [59,] 0.27735035 0.72264965
#>  [60,] 0.31131035 0.68868965
#>  [61,] 0.47862863 0.52137137
#>  [62,] 0.38292614 0.61707386
#>  [63,] 0.28174562 0.71825438
#>  [64,] 0.19042832 0.80957168
#>  [65,] 0.18864826 0.81135174
#>  [66,] 0.18962492 0.81037508
#>  [67,] 0.19153685 0.80846315
#>  [68,] 0.18446232 0.81553768
#>  [69,] 0.57441088 0.42558912
#>  [70,] 0.90527417 0.09472583
#>  [71,] 1.00000000 0.00000000
#>  [72,] 0.80957168 0.19042832
#>  [73,] 0.62144337 0.37855663
#>  [74,] 0.60259268 0.39740732
#>  [75,] 0.59573523 0.40426477
#>  [76,] 0.81135174 0.18864826
#>  [77,] 0.81030095 0.18969905
#>  [78,] 0.78896232 0.21103768
#>  [79,] 0.61601744 0.38398256
#>  [80,] 0.90607757 0.09392243
#>  [81,] 0.79603684 0.20396316
#>  [82,] 0.70733474 0.29266526
#>  [83,] 0.88995926 0.11004074
#>  [84,] 0.70026022 0.29973978
#>  [85,] 0.79603684 0.20396316
#>  [86,] 0.91129790 0.08870210
#>  [87,] 0.67354801 0.32645199
#>  [88,] 0.89900305 0.10099695
#>  [89,] 0.70937813 0.29062187
#>  [90,] 0.71557513 0.28442487
#>  [91,] 0.71898366 0.28101634
#>  [92,] 0.70857473 0.29142527
#>  [93,] 1.00000000 0.00000000
#>  [94,] 0.68705252 0.31294748
#>  [95,] 0.58527817 0.41472183
#>  [96,] 0.76827383 0.23172617
#>  [97,] 0.61028280 0.38971720
#>  [98,] 0.78663620 0.21336380
#>  [99,] 0.88763315 0.11236685
#> [100,] 0.80125717 0.19874283
#> [101,] 0.81559541 0.18440459
#> [102,] 0.78663620 0.21336380
#> [103,] 0.80330056 0.19669944
#> [104,] 0.72167298 0.27832702
#> [105,] 0.69325982 0.30674018
#> [106,] 0.62328116 0.37671884
#> [107,] 0.70293948 0.29706052
#> [108,] 0.79371073 0.20628927
#> [109,] 0.77759242 0.22240758
#> [110,] 0.88763315 0.11236685
#> [111,] 0.79603684 0.20396316
#> [112,] 0.89900305 0.10099695
#> [113,] 0.81030095 0.18969905
#> [114,] 0.80846315 0.19153685
#> [115,] 0.67957174 0.32042826
#> [116,] 0.67957174 0.32042826
#> [117,] 0.91129790 0.08870210
#> [118,] 0.61288793 0.38711207
#> [119,] 0.81030095 0.18969905
#> [120,] 0.80846315 0.19153685
#> [121,] 0.81559541 0.18440459
#> [122,] 0.89900305 0.10099695
#> [123,] 0.81030095 0.18969905
#> [124,] 0.81030095 0.18969905
#> [125,] 0.70033434 0.29966566
#> [126,] 0.90429751 0.09570249
#> [127,] 1.00000000 0.00000000
#> [128,] 1.00000000 0.00000000
#> [129,] 1.00000000 0.00000000
#> [130,] 0.80508062 0.19491938
#> [131,] 0.81037508 0.18962492
#> [132,] 1.00000000 0.00000000
#> [133,] 0.88995926 0.11004074
#> [134,] 0.88995926 0.11004074
#> [135,] 0.91129790 0.08870210
#> [136,] 1.00000000 0.00000000
#> [137,] 0.81553768 0.18446232
#> [138,] 0.90607757 0.09392243
#> [139,] 0.80330056 0.19669944
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.2213724 10.5704833  5.8649949  2.4934880  1.8384359  0.0000000  0.6590968 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  4.6866301  3.5053998  1.6948699  0.0000000  0.0000000  2.7196279 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  2.5407354  0.0000000  0.0000000  0.0000000  0.0000000  5.6536034 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.7313194  3.5721231  0.0000000  5.9932083  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  6.0448702  2.5487783  0.0000000  0.7407013  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  1.3645623  0.0000000  0.0000000  1.9738695  3.4937204  1.0384057  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.7566710  8.5972401  0.0000000  0.0000000  0.0000000  3.4344402  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  3.3217946  2.0023823  0.0000000  0.0000000  3.1617290  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  4.7754465 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 71 68 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V10        V49        V36        V33        V11        V28         V9 
#> 10.5704833  8.5972401  6.0448702  5.9932083  5.8649949  5.6536034  4.7754465 
#>        V17        V48        V31        V18        V45        V52        V55 
#>  4.6866301  3.7566710  3.5721231  3.5053998  3.4937204  3.4344402  3.3217946 
#>        V59        V21        V37        V23        V12        V56        V44 
#>  3.1617290  2.7196279  2.5487783  2.5407354  2.4934880  2.0023823  1.9738695 
#>        V13        V19        V41         V1        V46        V39        V30 
#>  1.8384359  1.6948699  1.3645623  1.2213724  1.0384057  0.7407013  0.7313194 
#>        V15        V14        V16         V2        V20        V22        V24 
#>  0.6590968  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V25        V26        V27        V29         V3        V32        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V38         V4        V40        V42        V43        V47 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V5        V50        V51        V53        V54        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V6        V60         V7         V8 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913