Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.55395683 0.44604317)  
#>   2) V11>=0.17885 91 23 M (0.74725275 0.25274725)  
#>     4) V6< 0.2171 81 13 M (0.83950617 0.16049383)  
#>       8) V29>=0.3601 71  6 M (0.91549296 0.08450704) *
#>       9) V29< 0.3601 10  3 R (0.30000000 0.70000000) *
#>     5) V6>=0.2171 10  0 R (0.00000000 1.00000000) *
#>   3) V11< 0.17885 48  9 R (0.18750000 0.81250000)  
#>     6) V21>=0.80695 11  3 M (0.72727273 0.27272727) *
#>     7) V21< 0.80695 37  1 R (0.02702703 0.97297297) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V45>=0.2781 35  0 M (1.00000000 0.00000000) *
#>    3) V45< 0.2781 104 36 R (0.34615385 0.65384615)  
#>      6) V37< 0.37555 64 30 M (0.53125000 0.46875000)  
#>       12) V48>=0.0718 33  7 M (0.78787879 0.21212121)  
#>         24) V13< 0.33705 26  1 M (0.96153846 0.03846154) *
#>         25) V13>=0.33705 7  1 R (0.14285714 0.85714286) *
#>       13) V48< 0.0718 31  8 R (0.25806452 0.74193548)  
#>         26) V15>=0.58075 7  0 M (1.00000000 0.00000000) *
#>         27) V15< 0.58075 24  1 R (0.04166667 0.95833333) *
#>      7) V37>=0.37555 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 R (0.44604317 0.55395683)  
#>   2) V12>=0.2107 81 28 M (0.65432099 0.34567901)  
#>     4) V39>=0.1625 63 13 M (0.79365079 0.20634921)  
#>       8) V47>=0.04165 53  3 M (0.94339623 0.05660377) *
#>       9) V47< 0.04165 10  0 R (0.00000000 1.00000000) *
#>     5) V39< 0.1625 18  3 R (0.16666667 0.83333333) *
#>   3) V12< 0.2107 58  9 R (0.15517241 0.84482759)  
#>     6) V4>=0.06685 7  1 M (0.85714286 0.14285714) *
#>     7) V4< 0.06685 51  3 R (0.05882353 0.94117647) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V23>=0.74575 72 19 M (0.73611111 0.26388889)  
#>      4) V36< 0.4924 54  4 M (0.92592593 0.07407407) *
#>      5) V36>=0.4924 18  3 R (0.16666667 0.83333333) *
#>    3) V23< 0.74575 67 18 R (0.26865672 0.73134328)  
#>      6) V54>=0.01715 7  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.01715 60 11 R (0.18333333 0.81666667)  
#>       14) V28>=0.89865 7  1 M (0.85714286 0.14285714) *
#>       15) V28< 0.89865 53  5 R (0.09433962 0.90566038) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V12>=0.2143 76 15 M (0.80263158 0.19736842)  
#>      4) V19>=0.26945 68  8 M (0.88235294 0.11764706) *
#>      5) V19< 0.26945 8  1 R (0.12500000 0.87500000) *
#>    3) V12< 0.2143 63 16 R (0.25396825 0.74603175)  
#>      6) V19>=0.8082 12  2 M (0.83333333 0.16666667) *
#>      7) V19< 0.8082 51  6 R (0.11764706 0.88235294)  
#>       14) V45>=0.39395 7  2 M (0.71428571 0.28571429) *
#>       15) V45< 0.39395 44  1 R (0.02272727 0.97727273) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V49>=0.04425 46  9 M (0.80434783 0.19565217)  
#>      4) V37< 0.4634 30  0 M (1.00000000 0.00000000) *
#>      5) V37>=0.4634 16  7 R (0.43750000 0.56250000) *
#>    3) V49< 0.04425 93 32 R (0.34408602 0.65591398)  
#>      6) V31< 0.35175 36 11 M (0.69444444 0.30555556)  
#>       12) V33>=0.19515 24  1 M (0.95833333 0.04166667) *
#>       13) V33< 0.19515 12  2 R (0.16666667 0.83333333) *
#>      7) V31>=0.35175 57  7 R (0.12280702 0.87719298)  
#>       14) V28>=0.9181 7  3 M (0.57142857 0.42857143) *
#>       15) V28< 0.9181 50  3 R (0.06000000 0.94000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V55< 0.00375 40  7 M (0.82500000 0.17500000)  
#>      4) V23< 0.9152 33  2 M (0.93939394 0.06060606) *
#>      5) V23>=0.9152 7  2 R (0.28571429 0.71428571) *
#>    3) V55>=0.00375 99 33 R (0.33333333 0.66666667)  
#>      6) V58>=0.00975 36 12 M (0.66666667 0.33333333)  
#>       12) V37< 0.5221 19  1 M (0.94736842 0.05263158) *
#>       13) V37>=0.5221 17  6 R (0.35294118 0.64705882) *
#>      7) V58< 0.00975 63  9 R (0.14285714 0.85714286)  
#>       14) V23>=0.836 9  4 M (0.55555556 0.44444444) *
#>       15) V23< 0.836 54  4 R (0.07407407 0.92592593) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V51>=0.01285 70 15 M (0.78571429 0.21428571)  
#>      4) V56>=0.00265 61  8 M (0.86885246 0.13114754) *
#>      5) V56< 0.00265 9  2 R (0.22222222 0.77777778) *
#>    3) V51< 0.01285 69 23 R (0.33333333 0.66666667)  
#>      6) V10>=0.20655 36 15 M (0.58333333 0.41666667)  
#>       12) V15< 0.56165 24  3 M (0.87500000 0.12500000) *
#>       13) V15>=0.56165 12  0 R (0.00000000 1.00000000) *
#>      7) V10< 0.20655 33  2 R (0.06060606 0.93939394) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V53>=0.0168 25  1 M (0.96000000 0.04000000) *
#>    3) V53< 0.0168 114 35 R (0.30701754 0.69298246)  
#>      6) V27>=0.82395 31  8 M (0.74193548 0.25806452)  
#>       12) V11>=0.19895 24  2 M (0.91666667 0.08333333) *
#>       13) V11< 0.19895 7  1 R (0.14285714 0.85714286) *
#>      7) V27< 0.82395 83 12 R (0.14457831 0.85542169)  
#>       14) V28< 0.35845 15  6 M (0.60000000 0.40000000) *
#>       15) V28>=0.35845 68  3 R (0.04411765 0.95588235) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 R (0.43884892 0.56115108)  
#>    2) V29>=0.7546 41  7 M (0.82926829 0.17073171)  
#>      4) V42>=0.13625 34  2 M (0.94117647 0.05882353) *
#>      5) V42< 0.13625 7  2 R (0.28571429 0.71428571) *
#>    3) V29< 0.7546 98 27 R (0.27551020 0.72448980)  
#>      6) V2>=0.0781 12  1 M (0.91666667 0.08333333) *
#>      7) V2< 0.0781 86 16 R (0.18604651 0.81395349)  
#>       14) V53>=0.0168 12  4 M (0.66666667 0.33333333) *
#>       15) V53< 0.0168 74  8 R (0.10810811 0.89189189) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.7375680 0.7640802 0.7193586 0.5908936 0.8071982 0.5671986
#>  [8] 0.7268610 0.8170560 0.5771470
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.5340592 5.5567412
#>   [2,] 2.1720298 4.9187706
#>   [3,] 1.3549738 5.7358266
#>   [4,] 2.8723786 4.2184219
#>   [5,] 2.3099960 4.7808044
#>   [6,] 1.9221724 5.1686280
#>   [7,] 2.8851442 4.2056562
#>   [8,] 2.1720298 4.9187706
#>   [9,] 1.5906374 5.5001631
#>  [10,] 2.6807603 4.4100401
#>  [11,] 0.0000000 7.0908004
#>  [12,] 2.8404226 4.2503778
#>  [13,] 2.8870524 4.2037480
#>  [14,] 1.5103001 5.5805003
#>  [15,] 1.5811362 5.5096642
#>  [16,] 2.1174504 4.9733500
#>  [17,] 2.6490334 4.4417670
#>  [18,] 2.0553226 5.0354778
#>  [19,] 1.2965057 5.7942947
#>  [20,] 1.5439170 5.5468834
#>  [21,] 1.1443457 5.9464547
#>  [22,] 2.0743324 5.0164680
#>  [23,] 1.4644290 5.6263715
#>  [24,] 1.9514798 5.1393206
#>  [25,] 0.7834391 6.3073613
#>  [26,] 2.2296588 4.8611417
#>  [27,] 1.3549738 5.7358266
#>  [28,] 2.7968574 4.2939430
#>  [29,] 1.1443457 5.9464547
#>  [30,] 2.1037039 4.9870965
#>  [31,] 2.0316276 5.0591728
#>  [32,] 1.3040080 5.7867924
#>  [33,] 0.5771470 6.5136534
#>  [34,] 2.0553226 5.0354778
#>  [35,] 2.0818348 5.0089656
#>  [36,] 2.8890330 4.2017674
#>  [37,] 2.0688468 5.0219536
#>  [38,] 0.5908936 6.4999068
#>  [39,] 1.3743328 5.7164677
#>  [40,] 2.1317710 4.9590294
#>  [41,] 0.0000000 7.0908004
#>  [42,] 1.4569266 5.6338738
#>  [43,] 1.8712067 5.2195938
#>  [44,] 0.5771470 6.5136534
#>  [45,] 0.0000000 7.0908004
#>  [46,] 0.0000000 7.0908004
#>  [47,] 0.5671986 6.5236018
#>  [48,] 0.8071982 6.2836022
#>  [49,] 0.5771470 6.5136534
#>  [50,] 0.7268610 6.3639394
#>  [51,] 1.3942030 5.6965974
#>  [52,] 1.3942030 5.6965974
#>  [53,] 0.7640802 6.3267202
#>  [54,] 1.8819136 5.2088868
#>  [55,] 0.0000000 7.0908004
#>  [56,] 0.0000000 7.0908004
#>  [57,] 0.5671986 6.5236018
#>  [58,] 0.7268610 6.3639394
#>  [59,] 2.8205524 4.2702480
#>  [60,] 1.5027978 5.5880026
#>  [61,] 2.1815310 4.9092694
#>  [62,] 2.1384130 4.9523875
#>  [63,] 1.3506378 5.7401626
#>  [64,] 1.2940596 5.7967408
#>  [65,] 1.5103001 5.5805003
#>  [66,] 0.7193586 6.3714418
#>  [67,] 1.9514798 5.1393206
#>  [68,] 5.6445808 1.4462196
#>  [69,] 4.9661340 2.1246664
#>  [70,] 4.8504347 2.2403657
#>  [71,] 5.7860338 1.3047666
#>  [72,] 5.6263715 1.4644290
#>  [73,] 5.9327082 1.1580922
#>  [74,] 5.6338738 1.4569266
#>  [75,] 5.0591728 2.0316276
#>  [76,] 4.9590294 2.1317710
#>  [77,] 5.1630156 1.9277848
#>  [78,] 7.0908004 0.0000000
#>  [79,] 5.7760854 1.3147150
#>  [80,] 4.7819540 2.3088464
#>  [81,] 6.2737444 0.8170560
#>  [82,] 6.3639394 0.7268610
#>  [83,] 5.1823745 1.9084259
#>  [84,] 7.0908004 0.0000000
#>  [85,] 4.9688872 2.1219132
#>  [86,] 5.5460342 1.5447662
#>  [87,] 5.5361764 1.5546240
#>  [88,] 4.9590294 2.1317710
#>  [89,] 5.7164036 1.3743968
#>  [90,] 4.8993476 2.1914528
#>  [91,] 4.3879420 2.7028584
#>  [92,] 4.9970449 2.0937555
#>  [93,] 6.3532324 0.7375680
#>  [94,] 5.1293987 1.9614017
#>  [95,] 5.4903053 1.6004951
#>  [96,] 6.3714418 0.7193586
#>  [97,] 6.3073613 0.7834391
#>  [98,] 4.4061513 2.6846491
#>  [99,] 7.0908004 0.0000000
#> [100,] 6.3073613 0.7834391
#> [101,] 4.8168178 2.2739826
#> [102,] 5.1057038 1.9850966
#> [103,] 4.4211116 2.6696888
#> [104,] 4.9424656 2.1483348
#> [105,] 4.9325172 2.1582832
#> [106,] 4.7720963 2.3187042
#> [107,] 4.3653186 2.7254819
#> [108,] 5.1255100 1.9652905
#> [109,] 5.1255100 1.9652905
#> [110,] 5.6927086 1.3980918
#> [111,] 5.1255100 1.9652905
#> [112,] 4.9230160 2.1677844
#> [113,] 5.6965974 1.3942030
#> [114,] 5.6965974 1.3942030
#> [115,] 3.9190269 3.1717735
#> [116,] 5.6445808 1.4462196
#> [117,] 5.0674338 2.0233667
#> [118,] 5.0567268 2.0340736
#> [119,] 6.3532324 0.7375680
#> [120,] 5.0567268 2.0340736
#> [121,] 5.6965974 1.3942030
#> [122,] 5.7064552 1.3843453
#> [123,] 5.7495732 1.3412272
#> [124,] 4.8168178 2.2739826
#> [125,] 6.5236018 0.5671986
#> [126,] 5.9327082 1.1580922
#> [127,] 5.7967408 1.2940596
#> [128,] 6.5236018 0.5671986
#> [129,] 5.7595216 1.3312788
#> [130,] 7.0908004 0.0000000
#> [131,] 5.7967408 1.2940596
#> [132,] 5.9327082 1.1580922
#> [133,] 6.5236018 0.5671986
#> [134,] 5.0227122 2.0680882
#> [135,] 5.7495732 1.3412272
#> [136,] 5.9227598 1.1680406
#> [137,] 5.3555611 1.7352393
#> [138,] 5.9227598 1.1680406
#> [139,] 5.0302146 2.0605859
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.21634500 0.78365500
#>   [2,] 0.30631659 0.69368341
#>   [3,] 0.19108898 0.80891102
#>   [4,] 0.40508524 0.59491476
#>   [5,] 0.32577366 0.67422634
#>   [6,] 0.27107975 0.72892025
#>   [7,] 0.40688555 0.59311445
#>   [8,] 0.30631659 0.69368341
#>   [9,] 0.22432409 0.77567591
#>  [10,] 0.37806173 0.62193827
#>  [11,] 0.00000000 1.00000000
#>  [12,] 0.40057856 0.59942144
#>  [13,] 0.40715466 0.59284534
#>  [14,] 0.21299431 0.78700569
#>  [15,] 0.22298416 0.77701584
#>  [16,] 0.29861938 0.70138062
#>  [17,] 0.37358736 0.62641264
#>  [18,] 0.28985762 0.71014238
#>  [19,] 0.18284335 0.81715665
#>  [20,] 0.21773522 0.78226478
#>  [21,] 0.16138456 0.83861544
#>  [22,] 0.29253854 0.70746146
#>  [23,] 0.20652520 0.79347480
#>  [24,] 0.27521291 0.72478709
#>  [25,] 0.11048670 0.88951330
#>  [26,] 0.31444388 0.68555612
#>  [27,] 0.19108898 0.80891102
#>  [28,] 0.39443465 0.60556535
#>  [29,] 0.16138456 0.83861544
#>  [30,] 0.29668074 0.70331926
#>  [31,] 0.28651597 0.71348403
#>  [32,] 0.18390139 0.81609861
#>  [33,] 0.08139378 0.91860622
#>  [34,] 0.28985762 0.71014238
#>  [35,] 0.29359659 0.70640341
#>  [36,] 0.40743398 0.59256602
#>  [37,] 0.29176492 0.70823508
#>  [38,] 0.08333243 0.91666757
#>  [39,] 0.19381913 0.80618087
#>  [40,] 0.30063898 0.69936102
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.20546716 0.79453284
#>  [43,] 0.26389216 0.73610784
#>  [44,] 0.08139378 0.91860622
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.07999078 0.92000922
#>  [48,] 0.11383739 0.88616261
#>  [49,] 0.08139378 0.91860622
#>  [50,] 0.10250761 0.89749239
#>  [51,] 0.19662139 0.80337861
#>  [52,] 0.19662139 0.80337861
#>  [53,] 0.10775655 0.89224345
#>  [54,] 0.26540215 0.73459785
#>  [55,] 0.00000000 1.00000000
#>  [56,] 0.00000000 1.00000000
#>  [57,] 0.07999078 0.92000922
#>  [58,] 0.10250761 0.89749239
#>  [59,] 0.39777630 0.60222370
#>  [60,] 0.21193627 0.78806373
#>  [61,] 0.30765652 0.69234348
#>  [62,] 0.30157568 0.69842432
#>  [63,] 0.19047748 0.80952252
#>  [64,] 0.18249838 0.81750162
#>  [65,] 0.21299431 0.78700569
#>  [66,] 0.10144957 0.89855043
#>  [67,] 0.27521291 0.72478709
#>  [68,] 0.79604283 0.20395717
#>  [69,] 0.70036297 0.29963703
#>  [70,] 0.68404614 0.31595386
#>  [71,] 0.81599163 0.18400837
#>  [72,] 0.79347480 0.20652520
#>  [73,] 0.83667680 0.16332320
#>  [74,] 0.79453284 0.20546716
#>  [75,] 0.71348403 0.28651597
#>  [76,] 0.69936102 0.30063898
#>  [77,] 0.72812874 0.27187126
#>  [78,] 1.00000000 0.00000000
#>  [79,] 0.81458863 0.18541137
#>  [80,] 0.67438847 0.32561153
#>  [81,] 0.88477239 0.11522761
#>  [82,] 0.89749239 0.10250761
#>  [83,] 0.73085889 0.26914111
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.70075124 0.29924876
#>  [86,] 0.78214502 0.21785498
#>  [87,] 0.78075480 0.21924520
#>  [88,] 0.69936102 0.30063898
#>  [89,] 0.80617183 0.19382817
#>  [90,] 0.69094422 0.30905578
#>  [91,] 0.61882182 0.38117818
#>  [92,] 0.70472227 0.29527773
#>  [93,] 0.89598241 0.10401759
#>  [94,] 0.72338783 0.27661217
#>  [95,] 0.77428569 0.22571431
#>  [96,] 0.89855043 0.10144957
#>  [97,] 0.88951330 0.11048670
#>  [98,] 0.62138984 0.37861016
#>  [99,] 1.00000000 0.00000000
#> [100,] 0.88951330 0.11048670
#> [101,] 0.67930523 0.32069477
#> [102,] 0.72004618 0.27995382
#> [103,] 0.62349965 0.37650035
#> [104,] 0.69702506 0.30297494
#> [105,] 0.69562206 0.30437794
#> [106,] 0.67299825 0.32700175
#> [107,] 0.61563128 0.38436872
#> [108,] 0.72283941 0.27716059
#> [109,] 0.72283941 0.27716059
#> [110,] 0.80283018 0.19716982
#> [111,] 0.72283941 0.27716059
#> [112,] 0.69428213 0.30571787
#> [113,] 0.80337861 0.19662139
#> [114,] 0.80337861 0.19662139
#> [115,] 0.55269175 0.44730825
#> [116,] 0.79604283 0.20395717
#> [117,] 0.71464905 0.28535095
#> [118,] 0.71313906 0.28686094
#> [119,] 0.89598241 0.10401759
#> [120,] 0.71313906 0.28686094
#> [121,] 0.80337861 0.19662139
#> [122,] 0.80476883 0.19523117
#> [123,] 0.81084967 0.18915033
#> [124,] 0.67930523 0.32069477
#> [125,] 0.92000922 0.07999078
#> [126,] 0.83667680 0.16332320
#> [127,] 0.81750162 0.18249838
#> [128,] 0.92000922 0.07999078
#> [129,] 0.81225267 0.18774733
#> [130,] 1.00000000 0.00000000
#> [131,] 0.81750162 0.18249838
#> [132,] 0.83667680 0.16332320
#> [133,] 0.92000922 0.07999078
#> [134,] 0.70834206 0.29165794
#> [135,] 0.81084967 0.18915033
#> [136,] 0.83527379 0.16472621
#> [137,] 0.75528302 0.24471698
#> [138,] 0.83527379 0.16472621
#> [139,] 0.70940010 0.29059990
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 2.0452041 6.1992799 7.5128733 1.6306355 0.0000000 4.8583034 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 3.2094341 1.9403989 0.0000000 1.9481705 0.0000000 4.7065435 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 3.9357050 4.1790353 4.6157127 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 3.4808358 0.0000000 2.4205972 0.0000000 0.0000000 3.3483206 5.1843136 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 2.5150562 1.7926689 0.0000000 0.0000000 0.8608797 0.0000000 0.0000000 5.9593806 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 3.4216072 1.9793046 3.1476299 0.0000000 0.0000000 3.0914337 0.0000000 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 5.3842125 1.7987766 2.3362790 1.4256250 0.0000000 2.1323448 0.0000000 2.9394380 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 72 67 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V12       V11       V45       V53       V37       V15       V23       V29 
#> 7.5128733 6.1992799 5.9593806 5.3842125 5.1843136 4.8583034 4.7065435 4.6157127 
#>       V28       V27       V31       V47       V36       V19       V49       V51 
#> 4.1790353 3.9357050 3.4808358 3.4216072 3.3483206 3.2094341 3.1476299 3.0914337 
#>        V6       V39       V33       V55       V58       V10       V48       V21 
#> 2.9394380 2.5150562 2.4205972 2.3362790 2.1323448 2.0452041 1.9793046 1.9481705 
#>        V2       V54        V4       V13       V56       V42        V1       V14 
#> 1.9403989 1.7987766 1.7926689 1.6306355 1.4256250 0.8608797 0.0000000 0.0000000 
#>       V16       V17       V18       V20       V22       V24       V25       V26 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>        V3       V30       V32       V34       V35       V38       V40       V41 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V43       V44       V46        V5       V50       V52       V57       V59 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1449275