Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 66 M (0.52517986 0.47482014)  
#>   2) V11>=0.21705 70 11 M (0.84285714 0.15714286)  
#>     4) V15< 0.66475 63  5 M (0.92063492 0.07936508) *
#>     5) V15>=0.66475 7  1 R (0.14285714 0.85714286) *
#>   3) V11< 0.21705 69 14 R (0.20289855 0.79710145)  
#>     6) V1>=0.03495 9  0 M (1.00000000 0.00000000) *
#>     7) V1< 0.03495 60  5 R (0.08333333 0.91666667) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V9>=0.1194 90 18 M (0.80000000 0.20000000)  
#>      4) V51>=0.01445 59  2 M (0.96610169 0.03389831) *
#>      5) V51< 0.01445 31 15 R (0.48387097 0.51612903)  
#>       10) V28>=0.7084 19  4 M (0.78947368 0.21052632) *
#>       11) V28< 0.7084 12  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.1194 49  8 R (0.16326531 0.83673469)  
#>      6) V19>=0.7815 8  1 M (0.87500000 0.12500000) *
#>      7) V19< 0.7815 41  1 R (0.02439024 0.97560976) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V12>=0.21485 90 28 M (0.68888889 0.31111111)  
#>      4) V27>=0.7701 43  1 M (0.97674419 0.02325581) *
#>      5) V27< 0.7701 47 20 R (0.42553191 0.57446809)  
#>       10) V52>=0.01415 18  2 M (0.88888889 0.11111111) *
#>       11) V52< 0.01415 29  4 R (0.13793103 0.86206897)  
#>         22) V2< 0.0264 7  3 M (0.57142857 0.42857143) *
#>         23) V2>=0.0264 22  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.21485 49  8 R (0.16326531 0.83673469)  
#>      6) V47>=0.18125 8  2 M (0.75000000 0.25000000) *
#>      7) V47< 0.18125 41  2 R (0.04878049 0.95121951) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V36< 0.4848 83 23 M (0.72289157 0.27710843)  
#>      4) V6>=0.07085 60  7 M (0.88333333 0.11666667)  
#>        8) V17< 0.86455 53  3 M (0.94339623 0.05660377) *
#>        9) V17>=0.86455 7  3 R (0.42857143 0.57142857) *
#>      5) V6< 0.07085 23  7 R (0.30434783 0.69565217)  
#>       10) V9>=0.14245 8  2 M (0.75000000 0.25000000) *
#>       11) V9< 0.14245 15  1 R (0.06666667 0.93333333) *
#>    3) V36>=0.4848 56 10 R (0.17857143 0.82142857)  
#>      6) V45>=0.33515 14  4 M (0.71428571 0.28571429) *
#>      7) V45< 0.33515 42  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V20>=0.51085 67 14 M (0.79104478 0.20895522)  
#>      4) V32< 0.38485 48  4 M (0.91666667 0.08333333) *
#>      5) V32>=0.38485 19  9 R (0.47368421 0.52631579) *
#>    3) V20< 0.51085 72 19 R (0.26388889 0.73611111)  
#>      6) V48>=0.08295 37 18 M (0.51351351 0.48648649)  
#>       12) V12>=0.20735 18  1 M (0.94444444 0.05555556) *
#>       13) V12< 0.20735 19  2 R (0.10526316 0.89473684) *
#>      7) V48< 0.08295 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V9>=0.1627 63 13 M (0.79365079 0.20634921)  
#>      4) V52>=0.0047 54  4 M (0.92592593 0.07407407) *
#>      5) V52< 0.0047 9  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.1627 76 23 R (0.30263158 0.69736842)  
#>      6) V48>=0.07165 30 10 M (0.66666667 0.33333333)  
#>       12) V36< 0.20965 17  1 M (0.94117647 0.05882353) *
#>       13) V36>=0.20965 13  4 R (0.30769231 0.69230769) *
#>      7) V48< 0.07165 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V31< 0.37765 47  7 M (0.85106383 0.14893617)  
#>      4) V30>=0.24885 39  2 M (0.94871795 0.05128205) *
#>      5) V30< 0.24885 8  3 R (0.37500000 0.62500000) *
#>    3) V31>=0.37765 92 28 R (0.30434783 0.69565217)  
#>      6) V18< 0.58935 52 26 M (0.50000000 0.50000000)  
#>       12) V12>=0.22995 25  2 M (0.92000000 0.08000000) *
#>       13) V12< 0.22995 27  3 R (0.11111111 0.88888889) *
#>      7) V18>=0.58935 40  2 R (0.05000000 0.95000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V5>=0.0392 108 35 M (0.67592593 0.32407407)  
#>      4) V36< 0.55155 79 15 M (0.81012658 0.18987342)  
#>        8) V44>=0.16295 48  2 M (0.95833333 0.04166667) *
#>        9) V44< 0.16295 31 13 M (0.58064516 0.41935484)  
#>         18) V54>=0.01475 12  0 M (1.00000000 0.00000000) *
#>         19) V54< 0.01475 19  6 R (0.31578947 0.68421053) *
#>      5) V36>=0.55155 29  9 R (0.31034483 0.68965517)  
#>       10) V44>=0.3649 11  2 M (0.81818182 0.18181818) *
#>       11) V44< 0.3649 18  0 R (0.00000000 1.00000000) *
#>    3) V5< 0.0392 31  4 R (0.12903226 0.87096774)  
#>      6) V24< 0.52605 7  3 M (0.57142857 0.42857143) *
#>      7) V24>=0.52605 24  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V31< 0.37765 44  2 M (0.95454545 0.04545455) *
#>    3) V31>=0.37765 95 38 R (0.40000000 0.60000000)  
#>      6) V37< 0.36985 47 12 M (0.74468085 0.25531915)  
#>       12) V52>=0.00725 38  3 M (0.92105263 0.07894737) *
#>       13) V52< 0.00725 9  0 R (0.00000000 1.00000000) *
#>      7) V37>=0.36985 48  3 R (0.06250000 0.93750000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 60 M (0.56834532 0.43165468)  
#>   2) V21>=0.51095 94 24 M (0.74468085 0.25531915)  
#>     4) V42>=0.08325 83 14 M (0.83132530 0.16867470)  
#>       8) V57< 0.01685 76  8 M (0.89473684 0.10526316) *
#>       9) V57>=0.01685 7  1 R (0.14285714 0.85714286) *
#>     5) V42< 0.08325 11  1 R (0.09090909 0.90909091) *
#>   3) V21< 0.51095 45  9 R (0.20000000 0.80000000)  
#>     6) V27>=0.8887 7  0 M (1.00000000 0.00000000) *
#>     7) V27< 0.8887 38  2 R (0.05263158 0.94736842) *
#> 
#> 
#> $weights
#>  [1] 0.6886628 0.7398314 0.7020366 0.8248952 0.7114462 0.7676095 0.7992524
#>  [8] 0.8713602 0.7270877 0.6492775
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.4790557 6.0024038
#>   [2,] 1.5733968 5.9080627
#>   [3,] 2.2533895 5.2280700
#>   [4,] 0.8713602 6.6100993
#>   [5,] 1.3763652 6.1050943
#>   [6,] 1.3763652 6.1050943
#>   [7,] 2.2320839 5.2493756
#>   [8,] 1.5012890 5.9801704
#>   [9,] 2.9591715 4.5222880
#>  [10,] 2.1305309 5.3509286
#>  [11,] 1.3763652 6.1050943
#>  [12,] 0.6492775 6.8321820
#>  [13,] 2.8075864 4.6738730
#>  [14,] 2.1677185 5.3137410
#>  [15,] 1.5269318 5.9545277
#>  [16,] 0.0000000 7.4814595
#>  [17,] 0.7270877 6.7543718
#>  [18,] 2.3977002 5.0837592
#>  [19,] 2.8169960 4.6644635
#>  [20,] 1.3607237 6.1207358
#>  [21,] 3.5068960 3.9745634
#>  [22,] 2.3323361 5.1491234
#>  [23,] 0.0000000 7.4814595
#>  [24,] 0.8713602 6.6100993
#>  [25,] 0.7270877 6.7543718
#>  [26,] 0.7270877 6.7543718
#>  [27,] 0.0000000 7.4814595
#>  [28,] 1.3513141 6.1301454
#>  [29,] 2.3982920 5.0831675
#>  [30,] 2.2871107 5.1943488
#>  [31,] 2.1505666 5.3308929
#>  [32,] 1.4485299 6.0329295
#>  [33,] 0.6492775 6.8321820
#>  [34,] 0.6492775 6.8321820
#>  [35,] 1.5206377 5.9608218
#>  [36,] 2.1161966 5.3652629
#>  [37,] 1.3607237 6.1207358
#>  [38,] 0.6492775 6.8321820
#>  [39,] 0.7398314 6.7416281
#>  [40,] 0.7398314 6.7416281
#>  [41,] 1.3891089 6.0923506
#>  [42,] 0.6492775 6.8321820
#>  [43,] 0.8713602 6.6100993
#>  [44,] 0.8713602 6.6100993
#>  [45,] 0.8713602 6.6100993
#>  [46,] 0.0000000 7.4814595
#>  [47,] 1.5263401 5.9551194
#>  [48,] 2.3512353 5.1302242
#>  [49,] 1.5263401 5.9551194
#>  [50,] 2.2918143 5.1896452
#>  [51,] 1.4669191 6.0145404
#>  [52,] 2.2634290 5.2180304
#>  [53,] 2.2634290 5.2180304
#>  [54,] 1.5363414 5.9451181
#>  [55,] 2.8460112 4.6354482
#>  [56,] 1.3607237 6.1207358
#>  [57,] 2.8453812 4.6360783
#>  [58,] 2.0493865 5.4320729
#>  [59,] 2.1189236 5.3625359
#>  [60,] 1.4696461 6.0118134
#>  [61,] 2.9548343 4.5266251
#>  [62,] 0.7676095 6.7138500
#>  [63,] 2.9938509 4.4876086
#>  [64,] 3.0526419 4.4288176
#>  [65,] 0.0000000 7.4814595
#>  [66,] 6.0923506 1.3891089
#>  [67,] 4.4338900 3.0475695
#>  [68,] 5.9935442 1.4879153
#>  [69,] 4.3782562 3.1032033
#>  [70,] 5.3231506 2.1583089
#>  [71,] 5.2625724 2.2188871
#>  [72,] 6.0118134 1.4696461
#>  [73,] 5.2259347 2.2555247
#>  [74,] 5.9145976 1.5668619
#>  [75,] 7.4814595 0.0000000
#>  [76,] 6.6822071 0.7992524
#>  [77,] 5.8108469 1.6706126
#>  [78,] 5.3542628 2.1271967
#>  [79,] 6.7927966 0.6886628
#>  [80,] 5.1875099 2.2939495
#>  [81,] 6.6822071 0.7992524
#>  [82,] 7.4814595 0.0000000
#>  [83,] 7.4814595 0.0000000
#>  [84,] 5.9679014 1.5135580
#>  [85,] 5.2193998 2.2620596
#>  [86,] 5.2658648 2.2155947
#>  [87,] 6.8321820 0.6492775
#>  [88,] 5.9080627 1.5733968
#>  [89,] 6.7543718 0.7270877
#>  [90,] 5.3030942 2.1783653
#>  [91,] 5.9214365 1.5600230
#>  [92,] 7.4814595 0.0000000
#>  [93,] 5.9707609 1.5106986
#>  [94,] 6.6100993 0.8713602
#>  [95,] 7.4814595 0.0000000
#>  [96,] 4.6328205 2.8486390
#>  [97,] 4.4039560 3.0775035
#>  [98,] 5.3214834 2.1599761
#>  [99,] 6.6100993 0.8713602
#> [100,] 5.9935442 1.4879153
#> [101,] 3.7681222 3.7133372
#> [102,] 6.0907600 1.3906995
#> [103,] 5.2915076 2.1899519
#> [104,] 6.7543718 0.7270877
#> [105,] 6.7543718 0.7270877
#> [106,] 6.7543718 0.7270877
#> [107,] 6.1050943 1.3763652
#> [108,] 5.2125610 2.2688985
#> [109,] 6.0301819 1.4512776
#> [110,] 6.0301819 1.4512776
#> [111,] 4.4965882 2.9848713
#> [112,] 5.2853558 2.1961037
#> [113,] 5.2002920 2.2811675
#> [114,] 5.9451181 1.5363414
#> [115,] 5.8702679 1.6111916
#> [116,] 6.0429256 1.4385338
#> [117,] 6.6100993 0.8713602
#> [118,] 5.1715655 2.3098940
#> [119,] 6.7700133 0.7114462
#> [120,] 6.7700133 0.7114462
#> [121,] 6.7700133 0.7114462
#> [122,] 7.4814595 0.0000000
#> [123,] 6.8321820 0.6492775
#> [124,] 6.8321820 0.6492775
#> [125,] 5.9608218 1.5206377
#> [126,] 5.9214365 1.5600230
#> [127,] 5.8702679 1.6111916
#> [128,] 6.7700133 0.7114462
#> [129,] 7.4814595 0.0000000
#> [130,] 7.4814595 0.0000000
#> [131,] 6.6565643 0.8248952
#> [132,] 7.4814595 0.0000000
#> [133,] 5.9608218 1.5206377
#> [134,] 7.4814595 0.0000000
#> [135,] 6.6100993 0.8713602
#> [136,] 7.4814595 0.0000000
#> [137,] 6.6100993 0.8713602
#> [138,] 6.6100993 0.8713602
#> [139,] 5.8986531 1.5828064
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.19769614 0.80230386
#>   [2,] 0.21030613 0.78969387
#>   [3,] 0.30119651 0.69880349
#>   [4,] 0.11646928 0.88353072
#>   [5,] 0.18397014 0.81602986
#>   [6,] 0.18397014 0.81602986
#>   [7,] 0.29834872 0.70165128
#>   [8,] 0.20066794 0.79933206
#>   [9,] 0.39553399 0.60446601
#>  [10,] 0.28477477 0.71522523
#>  [11,] 0.18397014 0.81602986
#>  [12,] 0.08678487 0.91321513
#>  [13,] 0.37527256 0.62472744
#>  [14,] 0.28974540 0.71025460
#>  [15,] 0.20409545 0.79590455
#>  [16,] 0.00000000 1.00000000
#>  [17,] 0.09718527 0.90281473
#>  [18,] 0.32048563 0.67951437
#>  [19,] 0.37653028 0.62346972
#>  [20,] 0.18187944 0.81812056
#>  [21,] 0.46874491 0.53125509
#>  [22,] 0.31174881 0.68825119
#>  [23,] 0.00000000 1.00000000
#>  [24,] 0.11646928 0.88353072
#>  [25,] 0.09718527 0.90281473
#>  [26,] 0.09718527 0.90281473
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.18062173 0.81937827
#>  [29,] 0.32056472 0.67943528
#>  [30,] 0.30570381 0.69429619
#>  [31,] 0.28745281 0.71254719
#>  [32,] 0.19361596 0.80638404
#>  [33,] 0.08678487 0.91321513
#>  [34,] 0.08678487 0.91321513
#>  [35,] 0.20325415 0.79674585
#>  [36,] 0.28285879 0.71714121
#>  [37,] 0.18187944 0.81812056
#>  [38,] 0.08678487 0.91321513
#>  [39,] 0.09888865 0.90111135
#>  [40,] 0.09888865 0.90111135
#>  [41,] 0.18567352 0.81432648
#>  [42,] 0.08678487 0.91321513
#>  [43,] 0.11646928 0.88353072
#>  [44,] 0.11646928 0.88353072
#>  [45,] 0.11646928 0.88353072
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.20401635 0.79598365
#>  [48,] 0.31427495 0.68572505
#>  [49,] 0.20401635 0.79598365
#>  [50,] 0.30633251 0.69366749
#>  [51,] 0.19607392 0.80392608
#>  [52,] 0.30253843 0.69746157
#>  [53,] 0.30253843 0.69746157
#>  [54,] 0.20535316 0.79464684
#>  [55,] 0.38040856 0.61959144
#>  [56,] 0.18187944 0.81812056
#>  [57,] 0.38032435 0.61967565
#>  [58,] 0.27392871 0.72607129
#>  [59,] 0.28322329 0.71677671
#>  [60,] 0.19643842 0.80356158
#>  [61,] 0.39495427 0.60504573
#>  [62,] 0.10260157 0.89739843
#>  [63,] 0.40016936 0.59983064
#>  [64,] 0.40802759 0.59197241
#>  [65,] 0.00000000 1.00000000
#>  [66,] 0.81432648 0.18567352
#>  [67,] 0.59265041 0.40734959
#>  [68,] 0.80111965 0.19888035
#>  [69,] 0.58521418 0.41478582
#>  [70,] 0.71151231 0.28848769
#>  [71,] 0.70341521 0.29658479
#>  [72,] 0.80356158 0.19643842
#>  [73,] 0.69851808 0.30148192
#>  [74,] 0.79056735 0.20943265
#>  [75,] 1.00000000 0.00000000
#>  [76,] 0.89316892 0.10683108
#>  [77,] 0.77669964 0.22330036
#>  [78,] 0.71567089 0.28432911
#>  [79,] 0.90795074 0.09204926
#>  [80,] 0.69338208 0.30661792
#>  [81,] 0.89316892 0.10683108
#>  [82,] 1.00000000 0.00000000
#>  [83,] 1.00000000 0.00000000
#>  [84,] 0.79769214 0.20230786
#>  [85,] 0.69764461 0.30235539
#>  [86,] 0.70385529 0.29614471
#>  [87,] 0.91321513 0.08678487
#>  [88,] 0.78969387 0.21030613
#>  [89,] 0.90281473 0.09718527
#>  [90,] 0.70883151 0.29116849
#>  [91,] 0.79148146 0.20851854
#>  [92,] 1.00000000 0.00000000
#>  [93,] 0.79807434 0.20192566
#>  [94,] 0.88353072 0.11646928
#>  [95,] 1.00000000 0.00000000
#>  [96,] 0.61924021 0.38075979
#>  [97,] 0.58864932 0.41135068
#>  [98,] 0.71128947 0.28871053
#>  [99,] 0.88353072 0.11646928
#> [100,] 0.80111965 0.19888035
#> [101,] 0.50366138 0.49633862
#> [102,] 0.81411388 0.18588612
#> [103,] 0.70728280 0.29271720
#> [104,] 0.90281473 0.09718527
#> [105,] 0.90281473 0.09718527
#> [106,] 0.90281473 0.09718527
#> [107,] 0.81602986 0.18397014
#> [108,] 0.69673050 0.30326950
#> [109,] 0.80601678 0.19398322
#> [110,] 0.80601678 0.19398322
#> [111,] 0.60103088 0.39896912
#> [112,] 0.70646052 0.29353948
#> [113,] 0.69509058 0.30490942
#> [114,] 0.79464684 0.20535316
#> [115,] 0.78464208 0.21535792
#> [116,] 0.80772016 0.19227984
#> [117,] 0.88353072 0.11646928
#> [118,] 0.69125088 0.30874912
#> [119,] 0.90490543 0.09509457
#> [120,] 0.90490543 0.09509457
#> [121,] 0.90490543 0.09509457
#> [122,] 1.00000000 0.00000000
#> [123,] 0.91321513 0.08678487
#> [124,] 0.91321513 0.08678487
#> [125,] 0.79674585 0.20325415
#> [126,] 0.79148146 0.20851854
#> [127,] 0.78464208 0.21535792
#> [128,] 0.90490543 0.09509457
#> [129,] 1.00000000 0.00000000
#> [130,] 1.00000000 0.00000000
#> [131,] 0.88974141 0.11025859
#> [132,] 1.00000000 0.00000000
#> [133,] 0.79674585 0.20325415
#> [134,] 1.00000000 0.00000000
#> [135,] 0.88353072 0.11646928
#> [136,] 1.00000000 0.00000000
#> [137,] 0.88353072 0.11646928
#> [138,] 0.88353072 0.11646928
#> [139,] 0.78843615 0.21156385
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 2.4135067 0.0000000 5.2229640 9.3652724 0.0000000 0.0000000 1.3987257 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.7204679 1.9501086 1.9096138 0.6487550 3.6566802 3.1240246 0.0000000 0.0000000 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.8217572 0.0000000 0.0000000 4.3884174 1.8074053 0.0000000 0.0000000 0.9307046 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 7.5439376 1.0127668 0.0000000 0.0000000 0.0000000 8.0253770 4.2822854 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 0.0000000 0.0000000 0.0000000 1.8424657 0.0000000 3.3701695 2.3550835 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 1.2314586 4.4853450 0.0000000 3.3454030 0.0000000 1.8633225 7.4409327 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 1.5989124 0.0000000 0.0000000 1.2538192 0.0000000 0.0000000 2.4502447 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 9.5400726 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V9       V12       V36       V31       V52       V11       V48       V27 
#> 9.5400726 9.3652724 8.0253770 7.5439376 7.4409327 5.2229640 4.4853450 4.3884174 
#>       V37       V20       V44        V5       V21        V6        V1       V45 
#> 4.2822854 3.6566802 3.3701695 3.3454030 3.1240246 2.4502447 2.4135067 2.3550835 
#>       V18       V19       V51       V42       V28       V54       V15       V57 
#> 1.9501086 1.9096138 1.8633225 1.8424657 1.8074053 1.5989124 1.3987257 1.2538192 
#>       V47       V32       V30       V24       V17        V2       V10       V13 
#> 1.2314586 1.0127668 0.9307046 0.8217572 0.7204679 0.6487550 0.0000000 0.0000000 
#>       V14       V16       V22       V23       V25       V26       V29        V3 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V33       V34       V35       V38       V39        V4       V40       V41 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V43       V46       V49       V50       V53       V55       V56       V58 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V59       V60        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913