Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V12>=0.1941 87 20 M (0.77011494 0.22988506)  
#>      4) V36< 0.4043 54  4 M (0.92592593 0.07407407) *
#>      5) V36>=0.4043 33 16 M (0.51515152 0.48484848)  
#>       10) V39>=0.3069 22  5 M (0.77272727 0.22727273)  
#>         20) V32< 0.5006 14  0 M (1.00000000 0.00000000) *
#>         21) V32>=0.5006 8  3 R (0.37500000 0.62500000) *
#>       11) V39< 0.3069 11  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.1941 52  8 R (0.15384615 0.84615385)  
#>      6) V4>=0.0522 10  3 M (0.70000000 0.30000000) *
#>      7) V4< 0.0522 42  1 R (0.02380952 0.97619048) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 61 M (0.56115108 0.43884892)  
#>   2) V49>=0.0424 78 14 M (0.82051282 0.17948718)  
#>     4) V9>=0.1352 61  3 M (0.95081967 0.04918033) *
#>     5) V9< 0.1352 17  6 R (0.35294118 0.64705882) *
#>   3) V49< 0.0424 61 14 R (0.22950820 0.77049180)  
#>     6) V27< 0.44115 9  0 M (1.00000000 0.00000000) *
#>     7) V27>=0.44115 52  5 R (0.09615385 0.90384615) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.4820144 0.5179856)  
#>    2) V49>=0.0581 45  6 M (0.8666667 0.1333333) *
#>    3) V49< 0.0581 94 28 R (0.2978723 0.7021277)  
#>      6) V11>=0.2044 38 16 M (0.5789474 0.4210526)  
#>       12) V52>=0.0093 19  2 M (0.8947368 0.1052632) *
#>       13) V52< 0.0093 19  5 R (0.2631579 0.7368421) *
#>      7) V11< 0.2044 56  6 R (0.1071429 0.8928571)  
#>       14) V19>=0.81725 10  4 M (0.6000000 0.4000000) *
#>       15) V19< 0.81725 46  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V10>=0.19155 65 14 M (0.78461538 0.21538462)  
#>      4) V18< 0.5017 46  1 M (0.97826087 0.02173913) *
#>      5) V18>=0.5017 19  6 R (0.31578947 0.68421053) *
#>    3) V10< 0.19155 74 17 R (0.22972973 0.77027027)  
#>      6) V28>=0.96135 8  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.96135 66  9 R (0.13636364 0.86363636)  
#>       14) V43>=0.34935 8  2 M (0.75000000 0.25000000) *
#>       15) V43< 0.34935 58  3 R (0.05172414 0.94827586) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V12>=0.2281 70 15 M (0.78571429 0.21428571)  
#>      4) V10>=0.0935 56  3 M (0.94642857 0.05357143) *
#>      5) V10< 0.0935 14  2 R (0.14285714 0.85714286) *
#>    3) V12< 0.2281 69 18 R (0.26086957 0.73913043)  
#>      6) V20>=0.51085 29 11 M (0.62068966 0.37931034)  
#>       12) V42>=0.1192 19  1 M (0.94736842 0.05263158) *
#>       13) V42< 0.1192 10  0 R (0.00000000 1.00000000) *
#>      7) V20< 0.51085 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V28>=0.80925 37  6 M (0.83783784 0.16216216)  
#>      4) V46>=0.08905 30  1 M (0.96666667 0.03333333) *
#>      5) V46< 0.08905 7  2 R (0.28571429 0.71428571) *
#>    3) V28< 0.80925 102 36 R (0.35294118 0.64705882)  
#>      6) V19>=0.88995 13  1 M (0.92307692 0.07692308) *
#>      7) V19< 0.88995 89 24 R (0.26966292 0.73033708)  
#>       14) V48>=0.08655 37 17 M (0.54054054 0.45945946)  
#>         28) V20>=0.29695 25  5 M (0.80000000 0.20000000)  
#>           56) V11>=0.17 16  0 M (1.00000000 0.00000000) *
#>           57) V11< 0.17 9  4 R (0.44444444 0.55555556) *
#>         29) V20< 0.29695 12  0 R (0.00000000 1.00000000) *
#>       15) V48< 0.08655 52  4 R (0.07692308 0.92307692) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V1>=0.0296 54 12 M (0.77777778 0.22222222)  
#>      4) V55>=0.0052 47  6 M (0.87234043 0.12765957) *
#>      5) V55< 0.0052 7  1 R (0.14285714 0.85714286) *
#>    3) V1< 0.0296 85 28 R (0.32941176 0.67058824)  
#>      6) V10>=0.1909 28 10 M (0.64285714 0.35714286)  
#>       12) V47>=0.071 18  0 M (1.00000000 0.00000000) *
#>       13) V47< 0.071 10  0 R (0.00000000 1.00000000) *
#>      7) V10< 0.1909 57 10 R (0.17543860 0.82456140)  
#>       14) V28>=0.96135 7  2 M (0.71428571 0.28571429) *
#>       15) V28< 0.96135 50  5 R (0.10000000 0.90000000)  
#>         30) V22>=0.9015 7  3 M (0.57142857 0.42857143) *
#>         31) V22< 0.9015 43  1 R (0.02325581 0.97674419) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V48>=0.07635 69 16 M (0.76811594 0.23188406)  
#>      4) V14< 0.427 55  3 M (0.94545455 0.05454545) *
#>      5) V14>=0.427 14  1 R (0.07142857 0.92857143) *
#>    3) V48< 0.07635 70 23 R (0.32857143 0.67142857)  
#>      6) V13>=0.27795 36 14 M (0.61111111 0.38888889)  
#>       12) V21>=0.612 25  3 M (0.88000000 0.12000000) *
#>       13) V21< 0.612 11  0 R (0.00000000 1.00000000) *
#>      7) V13< 0.27795 34  1 R (0.02941176 0.97058824) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V34< 0.6725 113 50 M (0.55752212 0.44247788)  
#>      4) V1>=0.0228 58 13 M (0.77586207 0.22413793)  
#>        8) V15< 0.56165 44  4 M (0.90909091 0.09090909) *
#>        9) V15>=0.56165 14  5 R (0.35714286 0.64285714) *
#>      5) V1< 0.0228 55 18 R (0.32727273 0.67272727)  
#>       10) V45>=0.19685 11  1 M (0.90909091 0.09090909) *
#>       11) V45< 0.19685 44  8 R (0.18181818 0.81818182)  
#>         22) V38>=0.3173 13  6 M (0.53846154 0.46153846) *
#>         23) V38< 0.3173 31  1 R (0.03225806 0.96774194) *
#>    3) V34>=0.6725 26  1 R (0.03846154 0.96153846) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V23>=0.70325 54 11 M (0.79629630 0.20370370)  
#>      4) V46>=0.0693 45  4 M (0.91111111 0.08888889)  
#>        8) V41< 0.3736 38  0 M (1.00000000 0.00000000) *
#>        9) V41>=0.3736 7  3 R (0.42857143 0.57142857) *
#>      5) V46< 0.0693 9  2 R (0.22222222 0.77777778) *
#>    3) V23< 0.70325 85 20 R (0.23529412 0.76470588)  
#>      6) V52>=0.01975 14  1 M (0.92857143 0.07142857) *
#>      7) V52< 0.01975 71  7 R (0.09859155 0.90140845)  
#>       14) V16< 0.1884 11  4 M (0.63636364 0.36363636) *
#>       15) V16>=0.1884 60  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.7537069 0.7649868 0.6634865 0.8290358 0.9102177 0.6830037
#>  [8] 0.8756814 0.6577396 0.7696305
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.2221716 5.4687563
#>   [2,] 2.2657463 5.4251815
#>   [3,] 3.0818469 4.6090810
#>   [4,] 2.1697790 5.5211488
#>   [5,] 1.6124749 6.0784529
#>   [6,] 2.9589651 4.7319628
#>   [7,] 3.2393344 4.4515934
#>   [8,] 0.0000000 7.6909278
#>   [9,] 1.4273701 6.2635578
#>  [10,] 1.4867754 6.2041525
#>  [11,] 1.4367105 6.2542173
#>  [12,] 0.8290358 6.8618921
#>  [13,] 2.1502618 5.5406660
#>  [14,] 1.4367105 6.2542173
#>  [15,] 2.8811074 4.8098204
#>  [16,] 2.2702145 5.4207133
#>  [17,] 1.5679572 6.1229706
#>  [18,] 2.4881563 5.2027715
#>  [19,] 2.3477294 5.3431984
#>  [20,] 2.9618942 4.7290336
#>  [21,] 1.6798482 6.0110797
#>  [22,] 1.5530697 6.1378582
#>  [23,] 2.4881563 5.2027715
#>  [24,] 1.5679572 6.1229706
#>  [25,] 0.7696305 6.9212973
#>  [26,] 1.5484259 6.1425019
#>  [27,] 0.0000000 7.6909278
#>  [28,] 1.4526342 6.2382936
#>  [29,] 0.0000000 7.6909278
#>  [30,] 0.6634865 7.0274414
#>  [31,] 0.6634865 7.0274414
#>  [32,] 0.6577396 7.0331883
#>  [33,] 2.9618942 4.7290336
#>  [34,] 2.3877209 5.3032070
#>  [35,] 2.2770262 5.4139016
#>  [36,] 2.1969075 5.4940204
#>  [37,] 1.5391679 6.1517600
#>  [38,] 1.4331170 6.2578109
#>  [39,] 2.3283156 5.3626122
#>  [40,] 0.6577396 7.0331883
#>  [41,] 0.0000000 7.6909278
#>  [42,] 1.5334210 6.1575069
#>  [43,] 0.9102177 6.7807102
#>  [44,] 2.2567078 5.4342200
#>  [45,] 1.6591205 6.0318073
#>  [46,] 0.6577396 7.0331883
#>  [47,] 0.6577396 7.0331883
#>  [48,] 0.6577396 7.0331883
#>  [49,] 0.6577396 7.0331883
#>  [50,] 1.3464902 6.3444377
#>  [51,] 2.3990188 5.2919091
#>  [52,] 2.1161207 5.5748072
#>  [53,] 1.4273701 6.2635578
#>  [54,] 0.0000000 7.6909278
#>  [55,] 0.7696305 6.9212973
#>  [56,] 0.0000000 7.6909278
#>  [57,] 0.7696305 6.9212973
#>  [58,] 1.4479905 6.2429374
#>  [59,] 2.3582081 5.3327197
#>  [60,] 1.4411787 6.2497491
#>  [61,] 0.6830037 7.0079242
#>  [62,] 2.5226926 5.1682353
#>  [63,] 2.9598723 4.7310555
#>  [64,] 1.5186936 6.1722342
#>  [65,] 2.2016973 5.4892305
#>  [66,] 2.2575090 5.4334188
#>  [67,] 1.4411787 6.2497491
#>  [68,] 0.7834391 6.9074887
#>  [69,] 1.4411787 6.2497491
#>  [70,] 5.2573728 2.4335550
#>  [71,] 6.8618921 0.8290358
#>  [72,] 5.4940204 2.1969075
#>  [73,] 4.8098204 2.8811074
#>  [74,] 7.0274414 0.6634865
#>  [75,] 5.5748072 2.1161207
#>  [76,] 4.5743691 3.1165587
#>  [77,] 6.1517600 1.5391679
#>  [78,] 6.0270033 1.6639245
#>  [79,] 7.0274414 0.6634865
#>  [80,] 5.5041040 2.1868238
#>  [81,] 6.1517600 1.5391679
#>  [82,] 6.1722342 1.5186936
#>  [83,] 6.1172237 1.5737041
#>  [84,] 4.7243899 2.9665380
#>  [85,] 5.4940204 2.1969075
#>  [86,] 6.3697018 1.3212261
#>  [87,] 5.4594841 2.2314437
#>  [88,] 5.3088224 2.3821054
#>  [89,] 6.9212973 0.7696305
#>  [90,] 6.1575069 1.5334210
#>  [91,] 7.0331883 0.6577396
#>  [92,] 5.9972710 1.6936568
#>  [93,] 5.5154019 2.1755259
#>  [94,] 7.6909278 0.0000000
#>  [95,] 6.9212973 0.7696305
#>  [96,] 5.9516744 1.7392534
#>  [97,] 6.0922616 1.5986663
#>  [98,] 7.0274414 0.6634865
#>  [99,] 7.0274414 0.6634865
#> [100,] 7.0274414 0.6634865
#> [101,] 6.2737345 1.4171933
#> [102,] 5.3683208 2.3226070
#> [103,] 6.1378582 1.5530697
#> [104,] 6.9074887 0.7834391
#> [105,] 5.4801186 2.2108092
#> [106,] 6.9074887 0.7834391
#> [107,] 6.9074887 0.7834391
#> [108,] 6.0922616 1.5986663
#> [109,] 4.8510082 2.8399197
#> [110,] 5.6000713 2.0908566
#> [111,] 5.1979676 2.4929603
#> [112,] 6.1537818 1.5371460
#> [113,] 6.0615396 1.6293883
#> [114,] 5.3785359 2.3123919
#> [115,] 6.2794814 1.4114464
#> [116,] 5.3439996 2.3469282
#> [117,] 5.2620166 2.4289113
#> [118,] 6.1722342 1.5186936
#> [119,] 7.6909278 0.0000000
#> [120,] 6.9212973 0.7696305
#> [121,] 6.0456159 1.6453119
#> [122,] 7.6909278 0.0000000
#> [123,] 7.0331883 0.6577396
#> [124,] 7.6909278 0.0000000
#> [125,] 5.5098509 2.1810769
#> [126,] 6.9372210 0.7537069
#> [127,] 6.2682015 1.4227263
#> [128,] 6.0157234 1.6752044
#> [129,] 7.6909278 0.0000000
#> [130,] 7.6909278 0.0000000
#> [131,] 7.0274414 0.6634865
#> [132,] 7.6909278 0.0000000
#> [133,] 7.6909278 0.0000000
#> [134,] 5.3280760 2.3628518
#> [135,] 7.0331883 0.6577396
#> [136,] 5.2573728 2.4335550
#> [137,] 6.0270033 1.6639245
#> [138,] 6.2429374 1.4479905
#> [139,] 5.1513219 2.5396059
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.28893413 0.71106587
#>   [2,] 0.29459987 0.70540013
#>   [3,] 0.40071198 0.59928802
#>   [4,] 0.28212188 0.71787812
#>   [5,] 0.20965935 0.79034065
#>   [6,] 0.38473448 0.61526552
#>   [7,] 0.42118903 0.57881097
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.18559140 0.81440860
#>  [10,] 0.19331547 0.80668453
#>  [11,] 0.18680588 0.81319412
#>  [12,] 0.10779399 0.89220601
#>  [13,] 0.27958419 0.72041581
#>  [14,] 0.18680588 0.81319412
#>  [15,] 0.37461117 0.62538883
#>  [16,] 0.29518083 0.70481917
#>  [17,] 0.20387101 0.79612899
#>  [18,] 0.32351835 0.67648165
#>  [19,] 0.30525958 0.69474042
#>  [20,] 0.38511533 0.61488467
#>  [21,] 0.21841944 0.78158056
#>  [22,] 0.20193528 0.79806472
#>  [23,] 0.32351835 0.67648165
#>  [24,] 0.20387101 0.79612899
#>  [25,] 0.10006992 0.89993008
#>  [26,] 0.20133148 0.79866852
#>  [27,] 0.00000000 1.00000000
#>  [28,] 0.18887633 0.81112367
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.08626872 0.91373128
#>  [31,] 0.08626872 0.91373128
#>  [32,] 0.08552149 0.91447851
#>  [33,] 0.38511533 0.61488467
#>  [34,] 0.31045940 0.68954060
#>  [35,] 0.29606652 0.70393348
#>  [36,] 0.28564921 0.71435079
#>  [37,] 0.20012772 0.79987228
#>  [38,] 0.18633863 0.81366137
#>  [39,] 0.30273533 0.69726467
#>  [40,] 0.08552149 0.91447851
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.19938049 0.80061951
#>  [43,] 0.11834952 0.88165048
#>  [44,] 0.29342465 0.70657535
#>  [45,] 0.21572437 0.78427563
#>  [46,] 0.08552149 0.91447851
#>  [47,] 0.08552149 0.91447851
#>  [48,] 0.08552149 0.91447851
#>  [49,] 0.08552149 0.91447851
#>  [50,] 0.17507513 0.82492487
#>  [51,] 0.31192839 0.68807161
#>  [52,] 0.27514504 0.72485496
#>  [53,] 0.18559140 0.81440860
#>  [54,] 0.00000000 1.00000000
#>  [55,] 0.10006992 0.89993008
#>  [56,] 0.00000000 1.00000000
#>  [57,] 0.10006992 0.89993008
#>  [58,] 0.18827253 0.81172747
#>  [59,] 0.30662206 0.69337794
#>  [60,] 0.18738685 0.81261315
#>  [61,] 0.08880641 0.91119359
#>  [62,] 0.32800887 0.67199113
#>  [63,] 0.38485244 0.61514756
#>  [64,] 0.19746559 0.80253441
#>  [65,] 0.28627200 0.71372800
#>  [66,] 0.29352883 0.70647117
#>  [67,] 0.18738685 0.81261315
#>  [68,] 0.10186536 0.89813464
#>  [69,] 0.18738685 0.81261315
#>  [70,] 0.68358109 0.31641891
#>  [71,] 0.89220601 0.10779399
#>  [72,] 0.71435079 0.28564921
#>  [73,] 0.62538883 0.37461117
#>  [74,] 0.91373128 0.08626872
#>  [75,] 0.72485496 0.27514504
#>  [76,] 0.59477468 0.40522532
#>  [77,] 0.79987228 0.20012772
#>  [78,] 0.78365101 0.21634899
#>  [79,] 0.91373128 0.08626872
#>  [80,] 0.71566190 0.28433810
#>  [81,] 0.79987228 0.20012772
#>  [82,] 0.80253441 0.19746559
#>  [83,] 0.79538176 0.20461824
#>  [84,] 0.61428088 0.38571912
#>  [85,] 0.71435079 0.28564921
#>  [86,] 0.82820980 0.17179020
#>  [87,] 0.70986027 0.29013973
#>  [88,] 0.69027073 0.30972927
#>  [89,] 0.89993008 0.10006992
#>  [90,] 0.80061951 0.19938049
#>  [91,] 0.91447851 0.08552149
#>  [92,] 0.77978511 0.22021489
#>  [93,] 0.71713089 0.28286911
#>  [94,] 1.00000000 0.00000000
#>  [95,] 0.89993008 0.10006992
#>  [96,] 0.77385649 0.22614351
#>  [97,] 0.79213610 0.20786390
#>  [98,] 0.91373128 0.08626872
#>  [99,] 0.91373128 0.08626872
#> [100,] 0.91373128 0.08626872
#> [101,] 0.81573181 0.18426819
#> [102,] 0.69800692 0.30199308
#> [103,] 0.79806472 0.20193528
#> [104,] 0.89813464 0.10186536
#> [105,] 0.71254324 0.28745676
#> [106,] 0.89813464 0.10186536
#> [107,] 0.89813464 0.10186536
#> [108,] 0.79213610 0.20786390
#> [109,] 0.63074421 0.36925579
#> [110,] 0.72813988 0.27186012
#> [111,] 0.67585702 0.32414298
#> [112,] 0.80013517 0.19986483
#> [113,] 0.78814152 0.21185848
#> [114,] 0.69933511 0.30066489
#> [115,] 0.81647904 0.18352096
#> [116,] 0.69484460 0.30515540
#> [117,] 0.68418488 0.31581512
#> [118,] 0.80253441 0.19746559
#> [119,] 1.00000000 0.00000000
#> [120,] 0.89993008 0.10006992
#> [121,] 0.78607108 0.21392892
#> [122,] 1.00000000 0.00000000
#> [123,] 0.91447851 0.08552149
#> [124,] 1.00000000 0.00000000
#> [125,] 0.71640913 0.28359087
#> [126,] 0.90200053 0.09799947
#> [127,] 0.81501239 0.18498761
#> [128,] 0.78218435 0.21781565
#> [129,] 1.00000000 0.00000000
#> [130,] 1.00000000 0.00000000
#> [131,] 0.91373128 0.08626872
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 0.69277415 0.30722585
#> [135,] 0.91447851 0.08552149
#> [136,] 0.68358109 0.31641891
#> [137,] 0.78365101 0.21634899
#> [138,] 0.81172747 0.18827253
#> [139,] 0.66979200 0.33020800
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 4.2568218 8.1665738 2.8170089 9.0686186 2.6667244 3.8422703 1.0953542 1.4911799 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 2.0151895 3.4331896 0.0000000 5.1947107 2.6662025 0.6359250 4.1168091 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 2.4314144 5.6226849 0.0000000 0.0000000 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 0.0000000 0.8018851 0.0000000 1.9278736 0.0000000 1.3936443 0.0000000 0.7945259 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 1.7656745 1.4891580 0.0000000 0.7645868 2.5091011 1.1706085 0.0000000 1.5757333 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 2.6428730 2.2598989 5.2026326 8.5146709 0.0000000 0.0000000 0.0000000 4.6831232 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 0.0000000 1.1397441 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 0.0000000 0.0000000 1.8435884 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 70 69 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V12       V49       V10       V28       V48       V20       V52        V1 
#> 9.0686186 8.5146709 8.1665738 5.6226849 5.2026326 5.1947107 4.6831232 4.2568218 
#>       V23       V14       V19       V11       V13       V21       V46       V42 
#> 4.1168091 3.8422703 3.4331896 2.8170089 2.6667244 2.6662025 2.6428730 2.5091011 
#>       V27       V47       V18       V34        V9       V39       V45       V16 
#> 2.4314144 2.2598989 2.0151895 1.9278736 1.8435884 1.7656745 1.5757333 1.4911799 
#>        V4       V36       V43       V55       V15       V32       V38       V41 
#> 1.4891580 1.3936443 1.1706085 1.1397441 1.0953542 0.8018851 0.7945259 0.7645868 
#>       V22       V17        V2       V24       V25       V26       V29        V3 
#> 0.6359250 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V30       V31       V33       V35       V37       V40       V44        V5 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V50       V51       V53       V54       V56       V57       V58       V59 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>        V6       V60        V7        V8 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696