Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.16895 84 20 M (0.76190476 0.23809524)  
#>      4) V51>=0.01265 62  5 M (0.91935484 0.08064516) *
#>      5) V51< 0.01265 22  7 R (0.31818182 0.68181818)  
#>       10) V57< 0.00665 8  2 M (0.75000000 0.25000000) *
#>       11) V57>=0.00665 14  1 R (0.07142857 0.92857143) *
#>    3) V11< 0.16895 55 11 R (0.20000000 0.80000000)  
#>      6) V60< 0.00155 9  1 M (0.88888889 0.11111111) *
#>      7) V60>=0.00155 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V37< 0.44055 82 18 M (0.78048780 0.21951220)  
#>      4) V48>=0.04945 61  5 M (0.91803279 0.08196721)  
#>        8) V39>=0.1282 54  0 M (1.00000000 0.00000000) *
#>        9) V39< 0.1282 7  2 R (0.28571429 0.71428571) *
#>      5) V48< 0.04945 21  8 R (0.38095238 0.61904762)  
#>       10) V26< 0.55885 7  0 M (1.00000000 0.00000000) *
#>       11) V26>=0.55885 14  1 R (0.07142857 0.92857143) *
#>    3) V37>=0.44055 57 11 R (0.19298246 0.80701754)  
#>      6) V31< 0.4166 13  4 M (0.69230769 0.30769231) *
#>      7) V31>=0.4166 44  2 R (0.04545455 0.95454545) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V11>=0.21685 75 26 M (0.65333333 0.34666667)  
#>      4) V17< 0.38145 42  1 M (0.97619048 0.02380952) *
#>      5) V17>=0.38145 33  8 R (0.24242424 0.75757576)  
#>       10) V40< 0.2255 7  1 M (0.85714286 0.14285714) *
#>       11) V40>=0.2255 26  2 R (0.07692308 0.92307692) *
#>    3) V11< 0.21685 64 13 R (0.20312500 0.79687500)  
#>      6) V5>=0.07105 22  9 M (0.59090909 0.40909091)  
#>       12) V36< 0.4891 15  2 M (0.86666667 0.13333333) *
#>       13) V36>=0.4891 7  0 R (0.00000000 1.00000000) *
#>      7) V5< 0.07105 42  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V12>=0.15395 99 36 M (0.63636364 0.36363636)  
#>      4) V27>=0.8167 34  2 M (0.94117647 0.05882353) *
#>      5) V27< 0.8167 65 31 R (0.47692308 0.52307692)  
#>       10) V4>=0.07535 24  2 M (0.91666667 0.08333333) *
#>       11) V4< 0.07535 41  9 R (0.21951220 0.78048780)  
#>         22) V2>=0.0584 11  4 M (0.63636364 0.36363636) *
#>         23) V2< 0.0584 30  2 R (0.06666667 0.93333333) *
#>    3) V12< 0.15395 40  4 R (0.10000000 0.90000000)  
#>      6) V6>=0.1232 7  3 M (0.57142857 0.42857143) *
#>      7) V6< 0.1232 33  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V12>=0.2076 94 32 M (0.65957447 0.34042553)  
#>      4) V52>=0.00885 61  8 M (0.86885246 0.13114754)  
#>        8) V36< 0.55005 50  1 M (0.98000000 0.02000000) *
#>        9) V36>=0.55005 11  4 R (0.36363636 0.63636364) *
#>      5) V52< 0.00885 33  9 R (0.27272727 0.72727273)  
#>       10) V58>=0.01035 7  0 M (1.00000000 0.00000000) *
#>       11) V58< 0.01035 26  2 R (0.07692308 0.92307692) *
#>    3) V12< 0.2076 45  5 R (0.11111111 0.88888889)  
#>      6) V9>=0.1717 9  4 M (0.55555556 0.44444444) *
#>      7) V9< 0.1717 36  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V55< 0.00685 49 13 M (0.73469388 0.26530612)  
#>      4) V13>=0.2108 30  2 M (0.93333333 0.06666667) *
#>      5) V13< 0.2108 19  8 R (0.42105263 0.57894737) *
#>    3) V55>=0.00685 90 26 R (0.28888889 0.71111111)  
#>      6) V55>=0.01235 30 10 M (0.66666667 0.33333333)  
#>       12) V52>=0.00975 18  1 M (0.94444444 0.05555556) *
#>       13) V52< 0.00975 12  3 R (0.25000000 0.75000000) *
#>      7) V55< 0.01235 60  6 R (0.10000000 0.90000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V48>=0.07485 68 16 M (0.76470588 0.23529412)  
#>      4) V15< 0.5141 59  7 M (0.88135593 0.11864407) *
#>      5) V15>=0.5141 9  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.07485 71 22 R (0.30985915 0.69014085)  
#>      6) V31< 0.36445 26  8 M (0.69230769 0.30769231)  
#>       12) V30>=0.24885 15  0 M (1.00000000 0.00000000) *
#>       13) V30< 0.24885 11  3 R (0.27272727 0.72727273) *
#>      7) V31>=0.36445 45  4 R (0.08888889 0.91111111) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.19615 87 22 M (0.74712644 0.25287356)  
#>      4) V27>=0.8191 39  0 M (1.00000000 0.00000000) *
#>      5) V27< 0.8191 48 22 M (0.54166667 0.45833333)  
#>       10) V51>=0.0136 30  6 M (0.80000000 0.20000000)  
#>         20) V53>=0.00675 23  1 M (0.95652174 0.04347826) *
#>         21) V53< 0.00675 7  2 R (0.28571429 0.71428571) *
#>       11) V51< 0.0136 18  2 R (0.11111111 0.88888889) *
#>    3) V11< 0.19615 52 10 R (0.19230769 0.80769231)  
#>      6) V51>=0.0225 9  3 M (0.66666667 0.33333333) *
#>      7) V51< 0.0225 43  4 R (0.09302326 0.90697674) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V23>=0.785 52  6 M (0.88461538 0.11538462)  
#>      4) V36< 0.4891 45  2 M (0.95555556 0.04444444) *
#>      5) V36>=0.4891 7  3 R (0.42857143 0.57142857) *
#>    3) V23< 0.785 87 31 R (0.35632184 0.64367816)  
#>      6) V28>=0.95425 13  1 M (0.92307692 0.07692308) *
#>      7) V28< 0.95425 74 19 R (0.25675676 0.74324324)  
#>       14) V55< 0.0034 9  1 M (0.88888889 0.11111111) *
#>       15) V55>=0.0034 65 11 R (0.16923077 0.83076923)  
#>         30) V51>=0.0136 25 10 R (0.40000000 0.60000000)  
#>           60) V1>=0.03975 7  0 M (1.00000000 0.00000000) *
#>           61) V1< 0.03975 18  3 R (0.16666667 0.83333333) *
#>         31) V51< 0.0136 40  1 R (0.02500000 0.97500000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V8>=0.1074 81 28 M (0.65432099 0.34567901)  
#>      4) V7< 0.16265 56  9 M (0.83928571 0.16071429)  
#>        8) V24>=0.3823 46  1 M (0.97826087 0.02173913) *
#>        9) V24< 0.3823 10  2 R (0.20000000 0.80000000) *
#>      5) V7>=0.16265 25  6 R (0.24000000 0.76000000) *
#>    3) V8< 0.1074 58 12 R (0.20689655 0.79310345)  
#>      6) V3>=0.06285 12  4 M (0.66666667 0.33333333) *
#>      7) V3< 0.06285 46  4 R (0.08695652 0.91304348)  
#>       14) V26>=0.8966 7  3 M (0.57142857 0.42857143) *
#>       15) V26< 0.8966 39  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.6665923 0.7405411 0.9310972 0.6834592 0.8628566 0.5601125 0.8264124
#>  [8] 0.7335093 0.8792799 0.4839654
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.6716383 5.6961876
#>   [2,] 2.1064284 5.2613975
#>   [3,] 2.7132968 4.6545291
#>   [4,] 2.1059847 5.2618412
#>   [5,] 2.6968735 4.6709524
#>   [6,] 2.8320311 4.5357948
#>   [7,] 2.8105252 4.5573007
#>   [8,] 0.8792799 6.4885461
#>   [9,] 0.7405411 6.6272848
#>  [10,] 0.0000000 7.3678259
#>  [11,] 0.8792799 6.4885461
#>  [12,] 2.0905926 5.2772333
#>  [13,] 1.4240004 5.9438256
#>  [14,] 2.0835608 5.2842651
#>  [15,] 2.3311987 5.0366272
#>  [16,] 2.9839220 4.3839039
#>  [17,] 0.4839654 6.8838605
#>  [18,] 2.2129081 5.1549178
#>  [19,] 1.9101640 5.4576619
#>  [20,] 2.0531172 5.3147087
#>  [21,] 2.1200342 5.2477917
#>  [22,] 0.8264124 6.5414135
#>  [23,] 1.4930046 5.8748213
#>  [24,] 0.8792799 6.4885461
#>  [25,] 0.8628566 6.5049693
#>  [26,] 1.5669535 5.8008724
#>  [27,] 1.7575095 5.6103164
#>  [28,] 1.3500515 6.0177744
#>  [29,] 2.0341630 5.3336629
#>  [30,] 0.7405411 6.6272848
#>  [31,] 0.0000000 7.3678259
#>  [32,] 0.4839654 6.8838605
#>  [33,] 2.3533303 5.0144956
#>  [34,] 2.5675262 4.8002997
#>  [35,] 1.9938370 5.3739889
#>  [36,] 1.4240004 5.9438256
#>  [37,] 1.3103778 6.0574481
#>  [38,] 1.3865249 5.9813010
#>  [39,] 0.0000000 7.3678259
#>  [40,] 1.1674247 6.2004012
#>  [41,] 0.4839654 6.8838605
#>  [42,] 0.4839654 6.8838605
#>  [43,] 0.4839654 6.8838605
#>  [44,] 0.4839654 6.8838605
#>  [45,] 0.0000000 7.3678259
#>  [46,] 0.8264124 6.5414135
#>  [47,] 0.8264124 6.5414135
#>  [48,] 1.5458721 5.8219538
#>  [49,] 1.3103778 6.0574481
#>  [50,] 1.3103778 6.0574481
#>  [51,] 1.3103778 6.0574481
#>  [52,] 1.3103778 6.0574481
#>  [53,] 2.9778017 4.3900242
#>  [54,] 1.3632453 6.0045806
#>  [55,] 2.1799335 5.1878924
#>  [56,] 1.8103770 5.5574489
#>  [57,] 1.4071334 5.9606925
#>  [58,] 1.6033977 5.7644282
#>  [59,] 1.5976894 5.7701365
#>  [60,] 0.6665923 6.7012336
#>  [61,] 2.0134143 5.3544116
#>  [62,] 1.5294489 5.8383770
#>  [63,] 2.1406427 5.2271832
#>  [64,] 1.3468221 6.0210038
#>  [65,] 1.5294489 5.8383770
#>  [66,] 1.3865249 5.9813010
#>  [67,] 5.2861710 2.0816549
#>  [68,] 5.9813010 1.3865249
#>  [69,] 5.1781682 2.1896577
#>  [70,] 3.5557053 3.8121206
#>  [71,] 4.1404804 3.2273455
#>  [72,] 5.0366272 2.3311987
#>  [73,] 5.2842651 2.0835608
#>  [74,] 4.5948053 2.7730206
#>  [75,] 5.7550367 1.6127892
#>  [76,] 5.8766162 1.4912097
#>  [77,] 5.9813010 1.3865249
#>  [78,] 6.5414135 0.8264124
#>  [79,] 6.4367287 0.9310972
#>  [80,] 6.5414135 0.8264124
#>  [81,] 6.4367287 0.9310972
#>  [82,] 7.3678259 0.0000000
#>  [83,] 5.1263509 2.2414750
#>  [84,] 5.2192540 2.1485719
#>  [85,] 6.0574481 1.3103778
#>  [86,] 6.1503511 1.2174748
#>  [87,] 6.3237479 1.0440780
#>  [88,] 5.2804627 2.0873632
#>  [89,] 5.1945915 2.1732344
#>  [90,] 5.3211214 2.0467045
#>  [91,] 5.2804627 2.0873632
#>  [92,] 7.3678259 0.0000000
#>  [93,] 6.1433193 1.2245066
#>  [94,] 5.5574489 1.8103770
#>  [95,] 6.5049693 0.8628566
#>  [96,] 7.3678259 0.0000000
#>  [97,] 6.3237479 1.0440780
#>  [98,] 5.2449743 2.1228516
#>  [99,] 4.8851483 2.4826776
#> [100,] 6.8838605 0.4839654
#> [101,] 7.3678259 0.0000000
#> [102,] 4.7273820 2.6404439
#> [103,] 4.4596477 2.9081782
#> [104,] 4.9067794 2.4610465
#> [105,] 4.8496975 2.5181284
#> [106,] 5.9677243 1.4001016
#> [107,] 5.2113474 2.1564785
#> [108,] 5.7644282 1.6033977
#> [109,] 5.5902386 1.7775873
#> [110,] 5.6256894 1.7421365
#> [111,] 4.7778758 2.5899501
#> [112,] 6.0574481 1.3103778
#> [113,] 5.2122222 2.1556037
#> [114,] 5.2100239 2.1578020
#> [115,] 5.2407599 2.1270660
#> [116,] 5.2192540 2.1485719
#> [117,] 4.8002997 2.5675262
#> [118,] 6.8077134 0.5601125
#> [119,] 6.4885461 0.8792799
#> [120,] 6.3237479 1.0440780
#> [121,] 6.8077134 0.5601125
#> [122,] 5.7480049 1.6198210
#> [123,] 7.3678259 0.0000000
#> [124,] 7.3678259 0.0000000
#> [125,] 6.3237479 1.0440780
#> [126,] 6.8077134 0.5601125
#> [127,] 6.0210038 1.3468221
#> [128,] 6.3237479 1.0440780
#> [129,] 5.2782645 2.0895614
#> [130,] 6.6272848 0.7405411
#> [131,] 6.6272848 0.7405411
#> [132,] 6.6272848 0.7405411
#> [133,] 6.5414135 0.8264124
#> [134,] 7.3678259 0.0000000
#> [135,] 6.8838605 0.4839654
#> [136,] 6.5049693 0.8628566
#> [137,] 6.8077134 0.5601125
#> [138,] 5.0655769 2.3022490
#> [139,] 5.2407599 2.1270660
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.22688352 0.77311648
#>   [2,] 0.28589551 0.71410449
#>   [3,] 0.36826288 0.63173712
#>   [4,] 0.28583529 0.71416471
#>   [5,] 0.36603383 0.63396617
#>   [6,] 0.38437813 0.61562187
#>   [7,] 0.38145924 0.61854076
#>   [8,] 0.11934048 0.88065952
#>   [9,] 0.10051013 0.89948987
#>  [10,] 0.00000000 1.00000000
#>  [11,] 0.11934048 0.88065952
#>  [12,] 0.28374620 0.71625380
#>  [13,] 0.19327280 0.80672720
#>  [14,] 0.28279181 0.71720819
#>  [15,] 0.31640253 0.68359747
#>  [16,] 0.40499356 0.59500644
#>  [17,] 0.06568633 0.93431367
#>  [18,] 0.30034750 0.69965250
#>  [19,] 0.25925749 0.74074251
#>  [20,] 0.27865984 0.72134016
#>  [21,] 0.28774217 0.71225783
#>  [22,] 0.11216502 0.88783498
#>  [23,] 0.20263842 0.79736158
#>  [24,] 0.11934048 0.88065952
#>  [25,] 0.11711143 0.88288857
#>  [26,] 0.21267515 0.78732485
#>  [27,] 0.23853842 0.76146158
#>  [28,] 0.18323607 0.81676393
#>  [29,] 0.27608728 0.72391272
#>  [30,] 0.10051013 0.89948987
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.06568633 0.93431367
#>  [33,] 0.31940634 0.68059366
#>  [34,] 0.34847814 0.65152186
#>  [35,] 0.27061403 0.72938597
#>  [36,] 0.19327280 0.80672720
#>  [37,] 0.17785135 0.82214865
#>  [38,] 0.18818644 0.81181356
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.15844900 0.84155100
#>  [41,] 0.06568633 0.93431367
#>  [42,] 0.06568633 0.93431367
#>  [43,] 0.06568633 0.93431367
#>  [44,] 0.06568633 0.93431367
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.11216502 0.88783498
#>  [47,] 0.11216502 0.88783498
#>  [48,] 0.20981388 0.79018612
#>  [49,] 0.17785135 0.82214865
#>  [50,] 0.17785135 0.82214865
#>  [51,] 0.17785135 0.82214865
#>  [52,] 0.17785135 0.82214865
#>  [53,] 0.40416287 0.59583713
#>  [54,] 0.18502681 0.81497319
#>  [55,] 0.29587202 0.70412798
#>  [56,] 0.24571387 0.75428613
#>  [57,] 0.19098353 0.80901647
#>  [58,] 0.21762156 0.78237844
#>  [59,] 0.21684679 0.78315321
#>  [60,] 0.09047340 0.90952660
#>  [61,] 0.27327116 0.72672884
#>  [62,] 0.20758483 0.79241517
#>  [63,] 0.29053926 0.70946074
#>  [64,] 0.18279776 0.81720224
#>  [65,] 0.20758483 0.79241517
#>  [66,] 0.18818644 0.81181356
#>  [67,] 0.71746688 0.28253312
#>  [68,] 0.81181356 0.18818644
#>  [69,] 0.70280817 0.29719183
#>  [70,] 0.48259899 0.51740101
#>  [71,] 0.56196772 0.43803228
#>  [72,] 0.68359747 0.31640253
#>  [73,] 0.71720819 0.28279181
#>  [74,] 0.62363108 0.37636892
#>  [75,] 0.78110379 0.21889621
#>  [76,] 0.79760519 0.20239481
#>  [77,] 0.81181356 0.18818644
#>  [78,] 0.88783498 0.11216502
#>  [79,] 0.87362661 0.12637339
#>  [80,] 0.88783498 0.11216502
#>  [81,] 0.87362661 0.12637339
#>  [82,] 1.00000000 0.00000000
#>  [83,] 0.69577525 0.30422475
#>  [84,] 0.70838454 0.29161546
#>  [85,] 0.82214865 0.17785135
#>  [86,] 0.83475794 0.16524206
#>  [87,] 0.85829226 0.14170774
#>  [88,] 0.71669211 0.28330789
#>  [89,] 0.70503722 0.29496278
#>  [90,] 0.72221052 0.27778948
#>  [91,] 0.71669211 0.28330789
#>  [92,] 1.00000000 0.00000000
#>  [93,] 0.83380354 0.16619646
#>  [94,] 0.75428613 0.24571387
#>  [95,] 0.88288857 0.11711143
#>  [96,] 1.00000000 0.00000000
#>  [97,] 0.85829226 0.14170774
#>  [98,] 0.71187544 0.28812456
#>  [99,] 0.66303797 0.33696203
#> [100,] 0.93431367 0.06568633
#> [101,] 1.00000000 0.00000000
#> [102,] 0.64162509 0.35837491
#> [103,] 0.60528679 0.39471321
#> [104,] 0.66597385 0.33402615
#> [105,] 0.65822639 0.34177361
#> [106,] 0.80997087 0.19002913
#> [107,] 0.70731142 0.29268858
#> [108,] 0.78237844 0.21762156
#> [109,] 0.75873652 0.24126348
#> [110,] 0.76354810 0.23645190
#> [111,] 0.64847838 0.35152162
#> [112,] 0.82214865 0.17785135
#> [113,] 0.70743015 0.29256985
#> [114,] 0.70713179 0.29286821
#> [115,] 0.71130344 0.28869656
#> [116,] 0.70838454 0.29161546
#> [117,] 0.65152186 0.34847814
#> [118,] 0.92397859 0.07602141
#> [119,] 0.88065952 0.11934048
#> [120,] 0.85829226 0.14170774
#> [121,] 0.92397859 0.07602141
#> [122,] 0.78014939 0.21985061
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 0.85829226 0.14170774
#> [126,] 0.92397859 0.07602141
#> [127,] 0.81720224 0.18279776
#> [128,] 0.85829226 0.14170774
#> [129,] 0.71639376 0.28360624
#> [130,] 0.89948987 0.10051013
#> [131,] 0.89948987 0.10051013
#> [132,] 0.89948987 0.10051013
#> [133,] 0.88783498 0.11216502
#> [134,] 1.00000000 0.00000000
#> [135,] 0.93431367 0.06568633
#> [136,] 0.88288857 0.11711143
#> [137,] 0.92397859 0.07602141
#> [138,] 0.68752668 0.31247332
#> [139,] 0.71130344 0.28869656
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "R" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.6660920  0.0000000 11.2938806  7.3088540  0.9257077  0.0000000  2.7138691 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  5.0155231  0.0000000  0.0000000  0.9665723  0.0000000  0.0000000 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  4.3240365  1.3035820  0.0000000  2.1209546  3.5752382  2.3369675 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.8380136  1.5017783  4.3679081  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  4.2079727  4.6531876  0.0000000  1.2675319  2.7224267  1.6923500 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  5.0216371  0.0000000  2.5411609  0.0000000  6.2403741  4.6078473  0.9589806 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  5.8084438  0.0000000  0.8459615  2.1952081  0.0000000  0.6977379 
#>        V60         V7         V8         V9 
#>  1.8429342  1.6264063  1.7727838  1.0380779 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 73 66 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V12        V51        V55        V48        V17        V37 
#> 11.2938806  7.3088540  6.2403741  5.8084438  5.0216371  5.0155231  4.6531876 
#>        V52        V31        V23        V36        V27         V4        V15 
#>  4.6078473  4.3679081  4.3240365  4.2079727  3.5752382  2.7224267  2.7138691 
#>         V5        V28        V58        V26        V60         V8        V40 
#>  2.5411609  2.3369675  2.1952081  2.1209546  1.8429342  1.7727838  1.6923500 
#>         V1         V7        V30        V24        V39         V9         V2 
#>  1.6660920  1.6264063  1.5017783  1.3035820  1.2675319  1.0380779  0.9665723 
#>        V53        V13        V57         V3         V6        V10        V14 
#>  0.9589806  0.9257077  0.8459615  0.8380136  0.6977379  0.0000000  0.0000000 
#>        V16        V18        V19        V20        V21        V22        V25 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V29        V32        V33        V34        V35        V38        V41 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V42        V43        V44        V45        V46        V47        V49 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V50        V54        V56        V59 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203