Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V51>=0.015 61 16 M (0.73770492 0.26229508)  
#>      4) V12>=0.2436 33  0 M (1.00000000 0.00000000) *
#>      5) V12< 0.2436 28 12 R (0.42857143 0.57142857)  
#>       10) V45>=0.252 11  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.252 17  1 R (0.05882353 0.94117647) *
#>    3) V51< 0.015 78 19 R (0.24358974 0.75641026)  
#>      6) V22>=0.9367 9  0 M (1.00000000 0.00000000) *
#>      7) V22< 0.9367 69 10 R (0.14492754 0.85507246)  
#>       14) V4>=0.05215 15  7 M (0.53333333 0.46666667) *
#>       15) V4< 0.05215 54  2 R (0.03703704 0.96296296) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V48>=0.0846 59 11 M (0.81355932 0.18644068)  
#>      4) V37< 0.4299 42  1 M (0.97619048 0.02380952) *
#>      5) V37>=0.4299 17  7 R (0.41176471 0.58823529) *
#>    3) V48< 0.0846 80 25 R (0.31250000 0.68750000)  
#>      6) V12>=0.16675 47 23 M (0.51063830 0.48936170)  
#>       12) V51>=0.0127 14  1 M (0.92857143 0.07142857) *
#>       13) V51< 0.0127 33 11 R (0.33333333 0.66666667)  
#>         26) V8< 0.05955 8  0 M (1.00000000 0.00000000) *
#>         27) V8>=0.05955 25  3 R (0.12000000 0.88000000) *
#>      7) V12< 0.16675 33  1 R (0.03030303 0.96969697) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V10>=0.12195 104 36 M (0.65384615 0.34615385)  
#>      4) V33>=0.2824 81 18 M (0.77777778 0.22222222)  
#>        8) V40< 0.60865 72 10 M (0.86111111 0.13888889)  
#>         16) V58>=0.0032 57  2 M (0.96491228 0.03508772) *
#>         17) V58< 0.0032 15  7 R (0.46666667 0.53333333) *
#>        9) V40>=0.60865 9  1 R (0.11111111 0.88888889) *
#>      5) V33< 0.2824 23  5 R (0.21739130 0.78260870)  
#>       10) V23>=0.86195 7  2 M (0.71428571 0.28571429) *
#>       11) V23< 0.86195 16  0 R (0.00000000 1.00000000) *
#>    3) V10< 0.12195 35  4 R (0.11428571 0.88571429)  
#>      6) V3>=0.04195 7  3 M (0.57142857 0.42857143) *
#>      7) V3< 0.04195 28  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 58 M (0.58273381 0.41726619)  
#>    2) V10>=0.18425 86 19 M (0.77906977 0.22093023)  
#>      4) V30>=0.24555 73  8 M (0.89041096 0.10958904)  
#>        8) V15< 0.6112 65  3 M (0.95384615 0.04615385) *
#>        9) V15>=0.6112 8  3 R (0.37500000 0.62500000) *
#>      5) V30< 0.24555 13  2 R (0.15384615 0.84615385) *
#>    3) V10< 0.18425 53 14 R (0.26415094 0.73584906)  
#>      6) V20>=0.87415 10  2 M (0.80000000 0.20000000) *
#>      7) V20< 0.87415 43  6 R (0.13953488 0.86046512)  
#>       14) V44>=0.3666 8  3 M (0.62500000 0.37500000) *
#>       15) V44< 0.3666 35  1 R (0.02857143 0.97142857) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V9>=0.1074 103 30 M (0.70873786 0.29126214)  
#>      4) V47>=0.06295 82 14 M (0.82926829 0.17073171)  
#>        8) V39>=0.16245 70  6 M (0.91428571 0.08571429) *
#>        9) V39< 0.16245 12  4 R (0.33333333 0.66666667) *
#>      5) V47< 0.06295 21  5 R (0.23809524 0.76190476)  
#>       10) V60>=0.0085 7  2 M (0.71428571 0.28571429) *
#>       11) V60< 0.0085 14  0 R (0.00000000 1.00000000) *
#>    3) V9< 0.1074 36  3 R (0.08333333 0.91666667) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V13>=0.25625 74 12 M (0.83783784 0.16216216)  
#>      4) V7< 0.2347 67  5 M (0.92537313 0.07462687) *
#>      5) V7>=0.2347 7  0 R (0.00000000 1.00000000) *
#>    3) V13< 0.25625 65 20 R (0.30769231 0.69230769)  
#>      6) V48>=0.0766 27  9 M (0.66666667 0.33333333)  
#>       12) V12>=0.10375 20  3 M (0.85000000 0.15000000) *
#>       13) V12< 0.10375 7  1 R (0.14285714 0.85714286) *
#>      7) V48< 0.0766 38  2 R (0.05263158 0.94736842) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V9>=0.11155 95 32 M (0.66315789 0.33684211)  
#>      4) V34< 0.692 82 21 M (0.74390244 0.25609756)  
#>        8) V32>=0.2487 65  9 M (0.86153846 0.13846154)  
#>         16) V10>=0.1453 55  3 M (0.94545455 0.05454545) *
#>         17) V10< 0.1453 10  4 R (0.40000000 0.60000000) *
#>        9) V32< 0.2487 17  5 R (0.29411765 0.70588235) *
#>      5) V34>=0.692 13  2 R (0.15384615 0.84615385) *
#>    3) V9< 0.11155 44  6 R (0.13636364 0.86363636) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V36< 0.4927 90 26 M (0.71111111 0.28888889)  
#>      4) V4>=0.03755 61  6 M (0.90163934 0.09836066)  
#>        8) V42>=0.0851 54  1 M (0.98148148 0.01851852) *
#>        9) V42< 0.0851 7  2 R (0.28571429 0.71428571) *
#>      5) V4< 0.03755 29  9 R (0.31034483 0.68965517)  
#>       10) V44>=0.1894 8  0 M (1.00000000 0.00000000) *
#>       11) V44< 0.1894 21  1 R (0.04761905 0.95238095) *
#>    3) V36>=0.4927 49 13 R (0.26530612 0.73469388)  
#>      6) V28>=0.9055 11  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.9055 38  2 R (0.05263158 0.94736842) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V51>=0.01305 75 20 M (0.73333333 0.26666667)  
#>      4) V11>=0.13035 56  4 M (0.92857143 0.07142857) *
#>      5) V11< 0.13035 19  3 R (0.15789474 0.84210526) *
#>    3) V51< 0.01305 64 18 R (0.28125000 0.71875000)  
#>      6) V55< 0.0062 21  7 M (0.66666667 0.33333333)  
#>       12) V37< 0.44565 13  1 M (0.92307692 0.07692308) *
#>       13) V37>=0.44565 8  2 R (0.25000000 0.75000000) *
#>      7) V55>=0.0062 43  4 R (0.09302326 0.90697674) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 M (0.50359712 0.49640288)  
#>    2) V45>=0.252 32  2 M (0.93750000 0.06250000) *
#>    3) V45< 0.252 107 40 R (0.37383178 0.62616822)  
#>      6) V37< 0.493 69 30 M (0.56521739 0.43478261)  
#>       12) V7< 0.1612 54 16 M (0.70370370 0.29629630)  
#>         24) V2>=0.0201 35  3 M (0.91428571 0.08571429) *
#>         25) V2< 0.0201 19  6 R (0.31578947 0.68421053) *
#>       13) V7>=0.1612 15  1 R (0.06666667 0.93333333) *
#>      7) V37>=0.493 38  1 R (0.02631579 0.97368421) *
#> 
#> 
#> $weights
#>  [1] 0.8090480 0.7878344 0.5377499 0.8407692 0.7336125 0.7620299 0.5234449
#>  [8] 0.8370839 0.7103999 0.6806368
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.9031893 4.3194202
#>   [2,] 3.0887986 4.1338108
#>   [3,] 2.6765115 4.5460979
#>   [4,] 1.9674573 5.2551521
#>   [5,] 2.1762792 5.0463302
#>   [6,] 0.0000000 7.2226094
#>   [7,] 0.7103999 6.5122095
#>   [8,] 1.4142493 5.8083601
#>   [9,] 0.5377499 6.6848595
#>  [10,] 2.3390035 4.8836059
#>  [11,] 0.7620299 6.4605796
#>  [12,] 1.3748339 5.8477756
#>  [13,] 2.3072823 4.9153271
#>  [14,] 2.1875671 5.0350423
#>  [15,] 2.2060423 5.0165672
#>  [16,] 0.8370839 6.3855255
#>  [17,] 0.0000000 7.2226094
#>  [18,] 1.3910368 5.8315727
#>  [19,] 2.1838819 5.0387276
#>  [20,] 2.3364115 4.8861979
#>  [21,] 2.2517147 4.9708947
#>  [22,] 0.8370839 6.3855255
#>  [23,] 2.2834359 4.9391736
#>  [24,] 1.3255843 5.8970251
#>  [25,] 2.0359843 5.1866252
#>  [26,] 0.5377499 6.6848595
#>  [27,] 1.3910368 5.8315727
#>  [28,] 0.8370839 6.3855255
#>  [29,] 0.0000000 7.2226094
#>  [30,] 2.3304540 4.8921554
#>  [31,] 2.1088278 5.1137816
#>  [32,] 1.5426605 5.6799490
#>  [33,] 2.2834359 4.9391736
#>  [34,] 1.2997798 5.9228296
#>  [35,] 0.7620299 6.4605796
#>  [36,] 0.7103999 6.5122095
#>  [37,] 0.7336125 6.4889970
#>  [38,] 2.2602642 4.9623452
#>  [39,] 0.0000000 7.2226094
#>  [40,] 0.5377499 6.6848595
#>  [41,] 1.3642140 5.8583954
#>  [42,] 0.0000000 7.2226094
#>  [43,] 0.7620299 6.4605796
#>  [44,] 1.2997798 5.9228296
#>  [45,] 0.7620299 6.4605796
#>  [46,] 0.0000000 7.2226094
#>  [47,] 2.3584899 4.8641195
#>  [48,] 0.6806368 6.5419726
#>  [49,] 0.7103999 6.5122095
#>  [50,] 1.5474839 5.6751256
#>  [51,] 0.0000000 7.2226094
#>  [52,] 1.3910368 5.8315727
#>  [53,] 0.7103999 6.5122095
#>  [54,] 0.8407692 6.3818403
#>  [55,] 0.0000000 7.2226094
#>  [56,] 0.0000000 7.2226094
#>  [57,] 1.6027990 5.6198104
#>  [58,] 2.9103931 4.3122163
#>  [59,] 3.2916959 3.9309135
#>  [60,] 2.2814778 4.9411316
#>  [61,] 1.2183868 6.0042226
#>  [62,] 0.7103999 6.5122095
#>  [63,] 1.4440124 5.7785970
#>  [64,] 1.2338448 5.9887646
#>  [65,] 2.0448918 5.1777177
#>  [66,] 2.1368638 5.0857457
#>  [67,] 0.7336125 6.4889970
#>  [68,] 2.0804104 5.1421990
#>  [69,] 0.7336125 6.4889970
#>  [70,] 4.2066543 3.0159551
#>  [71,] 6.1614146 1.0611948
#>  [72,] 5.1479954 2.0746140
#>  [73,] 5.3206454 1.9019640
#>  [74,] 5.7012034 1.5214060
#>  [75,] 5.5447563 1.6778531
#>  [76,] 5.1565041 2.0661053
#>  [77,] 4.9153271 2.3072823
#>  [78,] 3.7866809 3.4359285
#>  [79,] 4.4143830 2.8082265
#>  [80,] 5.1247829 2.0978265
#>  [81,] 4.4176867 2.8049227
#>  [82,] 5.8620806 1.3605288
#>  [83,] 5.0742462 2.1483632
#>  [84,] 7.2226094 0.0000000
#>  [85,] 5.2849152 1.9376942
#>  [86,] 6.0185277 1.2040817
#>  [87,] 6.5419726 0.6806368
#>  [88,] 5.9512470 1.2713624
#>  [89,] 5.7012034 1.5214060
#>  [90,] 6.6991645 0.5234449
#>  [91,] 5.1634535 2.0591559
#>  [92,] 5.1777585 2.0448509
#>  [93,] 5.1777585 2.0448509
#>  [94,] 4.8356612 2.3869482
#>  [95,] 5.5976911 1.6249183
#>  [96,] 5.0742462 2.1483632
#>  [97,] 4.3122163 2.9103931
#>  [98,] 6.6991645 0.5234449
#>  [99,] 5.8583954 1.3642140
#> [100,] 5.8583954 1.3642140
#> [101,] 5.9655521 1.2570574
#> [102,] 5.9887646 1.2338448
#> [103,] 7.2226094 0.0000000
#> [104,] 6.1614146 1.0611948
#> [105,] 5.0493474 2.1732620
#> [106,] 4.9391327 2.2834767
#> [107,] 5.0463302 2.1762792
#> [108,] 5.0205257 2.2020837
#> [109,] 5.7243751 1.4982343
#> [110,] 5.8970251 1.3255843
#> [111,] 5.5976911 1.6249183
#> [112,] 5.7243751 1.4982343
#> [113,] 5.1951747 2.0274348
#> [114,] 5.7329246 1.4896848
#> [115,] 6.6848595 0.5377499
#> [116,] 3.9295678 3.2930416
#> [117,] 4.4033817 2.8192277
#> [118,] 6.4347750 0.7878344
#> [119,] 6.5122095 0.7103999
#> [120,] 5.0879771 2.1346323
#> [121,] 6.4889970 0.7336125
#> [122,] 4.7357084 2.4869011
#> [123,] 7.2226094 0.0000000
#> [124,] 6.6848595 0.5377499
#> [125,] 6.6848595 0.5377499
#> [126,] 6.6848595 0.5377499
#> [127,] 6.4135614 0.8090480
#> [128,] 6.4889970 0.7336125
#> [129,] 7.2226094 0.0000000
#> [130,] 5.1247829 2.0978265
#> [131,] 5.5727923 1.6498171
#> [132,] 6.6848595 0.5377499
#> [133,] 4.8641195 2.3584899
#> [134,] 5.2706102 1.9519993
#> [135,] 5.9512470 1.2713624
#> [136,] 5.9512470 1.2713624
#> [137,] 6.3855255 0.8370839
#> [138,] 5.5764775 1.6461319
#> [139,] 5.9512470 1.2713624
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.40195850 0.59804150
#>   [2,] 0.42765688 0.57234312
#>   [3,] 0.37057404 0.62942596
#>   [4,] 0.27240256 0.72759744
#>   [5,] 0.30131481 0.69868519
#>   [6,] 0.00000000 1.00000000
#>   [7,] 0.09835779 0.90164221
#>   [8,] 0.19580864 0.80419136
#>   [9,] 0.07445369 0.92554631
#>  [10,] 0.32384466 0.67615534
#>  [11,] 0.10550617 0.89449383
#>  [12,] 0.19035141 0.80964859
#>  [13,] 0.31945273 0.68054727
#>  [14,] 0.30287767 0.69712233
#>  [15,] 0.30543563 0.69456437
#>  [16,] 0.11589772 0.88410228
#>  [17,] 0.00000000 1.00000000
#>  [18,] 0.19259477 0.80740523
#>  [19,] 0.30236743 0.69763257
#>  [20,] 0.32348579 0.67651421
#>  [21,] 0.31175917 0.68824083
#>  [22,] 0.11589772 0.88410228
#>  [23,] 0.31615109 0.68384891
#>  [24,] 0.18353261 0.81646739
#>  [25,] 0.28189040 0.71810960
#>  [26,] 0.07445369 0.92554631
#>  [27,] 0.19259477 0.80740523
#>  [28,] 0.11589772 0.88410228
#>  [29,] 0.00000000 1.00000000
#>  [30,] 0.32266095 0.67733905
#>  [31,] 0.29197589 0.70802411
#>  [32,] 0.21358769 0.78641231
#>  [33,] 0.31615109 0.68384891
#>  [34,] 0.17995986 0.82004014
#>  [35,] 0.10550617 0.89449383
#>  [36,] 0.09835779 0.90164221
#>  [37,] 0.10157167 0.89842833
#>  [38,] 0.31294288 0.68705712
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.07445369 0.92554631
#>  [41,] 0.18888105 0.81111895
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.10550617 0.89449383
#>  [44,] 0.17995986 0.82004014
#>  [45,] 0.10550617 0.89449383
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.32654264 0.67345736
#>  [48,] 0.09423697 0.90576303
#>  [49,] 0.09835779 0.90164221
#>  [50,] 0.21425551 0.78574449
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.19259477 0.80740523
#>  [53,] 0.09835779 0.90164221
#>  [54,] 0.11640795 0.88359205
#>  [55,] 0.00000000 1.00000000
#>  [56,] 0.00000000 1.00000000
#>  [57,] 0.22191412 0.77808588
#>  [58,] 0.40295590 0.59704410
#>  [59,] 0.45574885 0.54425115
#>  [60,] 0.31587999 0.68412001
#>  [61,] 0.16869066 0.83130934
#>  [62,] 0.09835779 0.90164221
#>  [63,] 0.19992946 0.80007054
#>  [64,] 0.17083089 0.82916911
#>  [65,] 0.28312368 0.71687632
#>  [66,] 0.29585758 0.70414242
#>  [67,] 0.10157167 0.89842833
#>  [68,] 0.28804138 0.71195862
#>  [69,] 0.10157167 0.89842833
#>  [70,] 0.58242860 0.41757140
#>  [71,] 0.85307321 0.14692679
#>  [72,] 0.71276116 0.28723884
#>  [73,] 0.73666526 0.26333474
#>  [74,] 0.78935508 0.21064492
#>  [75,] 0.76769433 0.23230567
#>  [76,] 0.71393921 0.28606079
#>  [77,] 0.68054727 0.31945273
#>  [78,] 0.52428155 0.47571845
#>  [79,] 0.61118949 0.38881051
#>  [80,] 0.70954729 0.29045271
#>  [81,] 0.61164691 0.38835309
#>  [82,] 0.81162919 0.18837081
#>  [83,] 0.70255027 0.29744973
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.73171826 0.26828174
#>  [86,] 0.83328993 0.16671007
#>  [87,] 0.90576303 0.09423697
#>  [88,] 0.82397464 0.17602536
#>  [89,] 0.78935508 0.21064492
#>  [90,] 0.92752690 0.07247310
#>  [91,] 0.71490138 0.28509862
#>  [92,] 0.71688198 0.28311802
#>  [93,] 0.71688198 0.28311802
#>  [94,] 0.66951720 0.33048280
#>  [95,] 0.77502337 0.22497663
#>  [96,] 0.70255027 0.29744973
#>  [97,] 0.59704410 0.40295590
#>  [98,] 0.92752690 0.07247310
#>  [99,] 0.81111895 0.18888105
#> [100,] 0.81111895 0.18888105
#> [101,] 0.82595524 0.17404476
#> [102,] 0.82916911 0.17083089
#> [103,] 1.00000000 0.00000000
#> [104,] 0.85307321 0.14692679
#> [105,] 0.69910293 0.30089707
#> [106,] 0.68384325 0.31615675
#> [107,] 0.69868519 0.30131481
#> [108,] 0.69511245 0.30488755
#> [109,] 0.79256329 0.20743671
#> [110,] 0.81646739 0.18353261
#> [111,] 0.77502337 0.22497663
#> [112,] 0.79256329 0.20743671
#> [113,] 0.71929331 0.28070669
#> [114,] 0.79374700 0.20625300
#> [115,] 0.92554631 0.07445369
#> [116,] 0.54406483 0.45593517
#> [117,] 0.60966632 0.39033368
#> [118,] 0.89092109 0.10907891
#> [119,] 0.90164221 0.09835779
#> [120,] 0.70445137 0.29554863
#> [121,] 0.89842833 0.10157167
#> [122,] 0.65567831 0.34432169
#> [123,] 1.00000000 0.00000000
#> [124,] 0.92554631 0.07445369
#> [125,] 0.92554631 0.07445369
#> [126,] 0.92554631 0.07445369
#> [127,] 0.88798398 0.11201602
#> [128,] 0.89842833 0.10157167
#> [129,] 1.00000000 0.00000000
#> [130,] 0.70954729 0.29045271
#> [131,] 0.77157603 0.22842397
#> [132,] 0.92554631 0.07445369
#> [133,] 0.67345736 0.32654264
#> [134,] 0.72973767 0.27026233
#> [135,] 0.82397464 0.17602536
#> [136,] 0.82397464 0.17602536
#> [137,] 0.88410228 0.11589772
#> [138,] 0.77208626 0.22791374
#> [139,] 0.82397464 0.17602536
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 7.2322468 3.4014393 5.3992150 4.2114096 0.0000000 1.1402985 0.0000000 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 0.0000000 0.0000000 1.7060871 1.6908463 0.0000000 2.6761603 0.7591883 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 0.0000000 3.6418497 0.0000000 0.5587626 2.8602333 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 0.0000000 1.2905205 1.7189774 1.1620544 0.0000000 2.9992844 5.3840930 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 1.4412795 4.5981785 1.3750798 0.0000000 1.4269011 0.0000000 3.6060756 5.7468161 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 0.0000000 2.4357174 6.3936663 0.0000000 0.0000000 0.0000000 8.2503944 0.0000000 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 0.0000000 0.0000000 1.8742595 0.0000000 0.0000000 0.9008157 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.9925501 4.1926110 2.1011223 6.8318664 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 70 69 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V51       V10        V9       V48       V45       V12       V37        V4 
#> 8.2503944 7.2322468 6.8318664 6.3936663 5.7468161 5.3992150 5.3840930 4.5981785 
#>       V13        V7       V28       V44       V11       V36       V30       V22 
#> 4.2114096 4.1926110 3.6418497 3.6060756 3.4014393 2.9992844 2.8602333 2.6761603 
#>       V47        V8       V55       V33        V2       V20       V39       V42 
#> 2.4357174 2.1011223 1.8742595 1.7189774 1.7060871 1.6908463 1.4412795 1.4269011 
#>       V40       V32       V34       V15       V60       V58       V23        V3 
#> 1.3750798 1.2905205 1.1620544 1.1402985 0.9925501 0.9008157 0.7591883 0.5587626 
#>        V1       V14       V16       V17       V18       V19       V21       V24 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V25       V26       V27       V29       V31       V35       V38       V41 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V43       V46       V49        V5       V50       V52       V53       V54 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V56       V57       V59        V6 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623