Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.53237410 0.46762590)  
#>    2) V11>=0.21765 79 16 M (0.79746835 0.20253165)  
#>      4) V28>=0.59025 53  1 M (0.98113208 0.01886792) *
#>      5) V28< 0.59025 26 11 R (0.42307692 0.57692308)  
#>       10) V51>=0.0136 13  2 M (0.84615385 0.15384615) *
#>       11) V51< 0.0136 13  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.21765 60 11 R (0.18333333 0.81666667)  
#>      6) V21>=0.79625 11  2 M (0.81818182 0.18181818) *
#>      7) V21< 0.79625 49  2 R (0.04081633 0.95918367) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V2>=0.05965 31  2 M (0.93548387 0.06451613) *
#>    3) V2< 0.05965 108 46 R (0.42592593 0.57407407)  
#>      6) V35< 0.2444 45 12 M (0.73333333 0.26666667)  
#>       12) V14< 0.4977 37  5 M (0.86486486 0.13513514)  
#>         24) V11>=0.15315 30  0 M (1.00000000 0.00000000) *
#>         25) V11< 0.15315 7  2 R (0.28571429 0.71428571) *
#>       13) V14>=0.4977 8  1 R (0.12500000 0.87500000) *
#>      7) V35>=0.2444 63 13 R (0.20634921 0.79365079)  
#>       14) V28< 0.5405 17  8 M (0.52941176 0.47058824) *
#>       15) V28>=0.5405 46  4 R (0.08695652 0.91304348) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V12>=0.22505 85 25 M (0.70588235 0.29411765)  
#>      4) V28>=0.6071 49  3 M (0.93877551 0.06122449) *
#>      5) V28< 0.6071 36 14 R (0.38888889 0.61111111)  
#>       10) V59>=0.0102 12  1 M (0.91666667 0.08333333) *
#>       11) V59< 0.0102 24  3 R (0.12500000 0.87500000) *
#>    3) V12< 0.22505 54  9 R (0.16666667 0.83333333)  
#>      6) V20>=0.8255 9  2 M (0.77777778 0.22222222) *
#>      7) V20< 0.8255 45  2 R (0.04444444 0.95555556) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V48>=0.04695 102 38 M (0.62745098 0.37254902)  
#>      4) V37< 0.3836 55  6 M (0.89090909 0.10909091) *
#>      5) V37>=0.3836 47 15 R (0.31914894 0.68085106)  
#>       10) V45>=0.26515 11  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.26515 36  4 R (0.11111111 0.88888889)  
#>         22) V54>=0.016 7  3 M (0.57142857 0.42857143) *
#>         23) V54< 0.016 29  0 R (0.00000000 1.00000000) *
#>    3) V48< 0.04695 37  1 R (0.02702703 0.97297297) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 51 R (0.36690647 0.63309353)  
#>    2) V43>=0.24725 49 14 M (0.71428571 0.28571429)  
#>      4) V33< 0.5845 33  2 M (0.93939394 0.06060606) *
#>      5) V33>=0.5845 16  4 R (0.25000000 0.75000000) *
#>    3) V43< 0.24725 90 16 R (0.17777778 0.82222222)  
#>      6) V35< 0.244 29 14 R (0.48275862 0.51724138)  
#>       12) V26< 0.55885 10  0 M (1.00000000 0.00000000) *
#>       13) V26>=0.55885 19  4 R (0.21052632 0.78947368) *
#>      7) V35>=0.244 61  2 R (0.03278689 0.96721311) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 62 M (0.5539568 0.4460432)  
#>   2) V52>=0.007 106 34 M (0.6792453 0.3207547)  
#>     4) V20< 0.92755 90 21 M (0.7666667 0.2333333)  
#>       8) V39>=0.1632 74 10 M (0.8648649 0.1351351) *
#>       9) V39< 0.1632 16  5 R (0.3125000 0.6875000) *
#>     5) V20>=0.92755 16  3 R (0.1875000 0.8125000) *
#>   3) V52< 0.007 33  5 R (0.1515152 0.8484848)  
#>     6) V54>=0.01245 8  3 M (0.6250000 0.3750000) *
#>     7) V54< 0.01245 25  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V36< 0.40435 86 26 M (0.69767442 0.30232558)  
#>      4) V43>=0.2164 35  0 M (1.00000000 0.00000000) *
#>      5) V43< 0.2164 51 25 R (0.49019608 0.50980392)  
#>       10) V28>=0.958 11  0 M (1.00000000 0.00000000) *
#>       11) V28< 0.958 40 14 R (0.35000000 0.65000000)  
#>         22) V8>=0.11625 18  5 M (0.72222222 0.27777778) *
#>         23) V8< 0.11625 22  1 R (0.04545455 0.95454545) *
#>    3) V36>=0.40435 53 11 R (0.20754717 0.79245283)  
#>      6) V46>=0.2041 10  1 M (0.90000000 0.10000000) *
#>      7) V46< 0.2041 43  2 R (0.04651163 0.95348837) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V11>=0.1579 99 29 M (0.70707071 0.29292929)  
#>      4) V6< 0.21915 91 21 M (0.76923077 0.23076923)  
#>        8) V21>=0.56125 66  7 M (0.89393939 0.10606061)  
#>         16) V23>=0.7701 45  0 M (1.00000000 0.00000000) *
#>         17) V23< 0.7701 21  7 M (0.66666667 0.33333333)  
#>           34) V51>=0.01275 13  0 M (1.00000000 0.00000000) *
#>           35) V51< 0.01275 8  1 R (0.12500000 0.87500000) *
#>        9) V21< 0.56125 25 11 R (0.44000000 0.56000000)  
#>         18) V31>=0.60065 10  0 M (1.00000000 0.00000000) *
#>         19) V31< 0.60065 15  1 R (0.06666667 0.93333333) *
#>      5) V6>=0.21915 8  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1579 40  7 R (0.17500000 0.82500000)  
#>      6) V5>=0.0695 9  2 M (0.77777778 0.22222222) *
#>      7) V5< 0.0695 31  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V11>=0.21705 76 23 M (0.69736842 0.30263158)  
#>      4) V37< 0.5718 51  3 M (0.94117647 0.05882353) *
#>      5) V37>=0.5718 25  5 R (0.20000000 0.80000000)  
#>       10) V17< 0.37845 7  2 M (0.71428571 0.28571429) *
#>       11) V17>=0.37845 18  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.21705 63 13 R (0.20634921 0.79365079)  
#>      6) V2>=0.05 8  1 M (0.87500000 0.12500000) *
#>      7) V2< 0.05 55  6 R (0.10909091 0.89090909) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V49>=0.04425 51 12 M (0.76470588 0.23529412)  
#>      4) V36< 0.48855 39  3 M (0.92307692 0.07692308) *
#>      5) V36>=0.48855 12  3 R (0.25000000 0.75000000) *
#>    3) V49< 0.04425 88 27 R (0.30681818 0.69318182)  
#>      6) V21>=0.6496 43 19 M (0.55813953 0.44186047)  
#>       12) V6< 0.1521 28  4 M (0.85714286 0.14285714) *
#>       13) V6>=0.1521 15  0 R (0.00000000 1.00000000) *
#>      7) V21< 0.6496 45  3 R (0.06666667 0.93333333) *
#> 
#> 
#> $weights
#>  [1] 0.7834391 0.8206519 0.7245050 0.6341746 0.6338552 0.5235666 0.8593810
#>  [8] 1.1192847 0.7953581 0.6959910
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.2767065 5.3135008
#>   [2,] 2.5869545 5.0032528
#>   [3,] 2.8908917 4.6993156
#>   [4,] 3.5337651 4.0564422
#>   [5,] 2.2889137 5.3012936
#>   [6,] 2.9992933 4.5909140
#>   [7,] 2.8381098 4.7520975
#>   [8,] 0.0000000 7.5902073
#>   [9,] 1.2195575 6.3706498
#>  [10,] 1.8822462 5.7079611
#>  [11,] 2.4381783 5.1520290
#>  [12,] 1.6428513 5.9473560
#>  [13,] 3.1626833 4.4275240
#>  [14,] 1.3442185 6.2459888
#>  [15,] 3.1880082 4.4021991
#>  [16,] 1.6160100 5.9741973
#>  [17,] 0.0000000 7.5902073
#>  [18,] 2.5741113 5.0160960
#>  [19,] 0.7245050 6.8657023
#>  [20,] 1.6160100 5.9741973
#>  [21,] 1.3583602 6.2318471
#>  [22,] 2.8381098 4.7520975
#>  [23,] 1.9408609 5.6493464
#>  [24,] 1.9440626 5.6461447
#>  [25,] 2.9617448 4.6284625
#>  [26,] 0.5235666 7.0666408
#>  [27,] 0.6338552 6.9563521
#>  [28,] 1.6428513 5.9473560
#>  [29,] 0.5235666 7.0666408
#>  [30,] 0.5235666 7.0666408
#>  [31,] 2.3673563 5.2228510
#>  [32,] 1.3442185 6.2459888
#>  [33,] 1.6160100 5.9741973
#>  [34,] 2.3142076 5.2759997
#>  [35,] 1.4545072 6.1357001
#>  [36,] 1.9146428 5.6755645
#>  [37,] 0.0000000 7.5902073
#>  [38,] 0.6341746 6.9560327
#>  [39,] 0.6341746 6.9560327
#>  [40,] 1.4548265 6.1353808
#>  [41,] 0.0000000 7.5902073
#>  [42,] 0.0000000 7.5902073
#>  [43,] 0.0000000 7.5902073
#>  [44,] 1.2195575 6.3706498
#>  [45,] 0.6959910 6.8942163
#>  [46,] 0.5235666 7.0666408
#>  [47,] 0.0000000 7.5902073
#>  [48,] 0.8206519 6.7695554
#>  [49,] 1.1577412 6.4324661
#>  [50,] 0.0000000 7.5902073
#>  [51,] 0.8593810 6.7308263
#>  [52,] 1.3301656 6.2600417
#>  [53,] 1.4794301 6.1107772
#>  [54,] 2.6272289 4.9629784
#>  [55,] 2.5987149 4.9914925
#>  [56,] 2.9849047 4.6053026
#>  [57,] 3.6334679 3.9567394
#>  [58,] 2.1255237 5.4646836
#>  [59,] 2.2382657 5.3519416
#>  [60,] 0.0000000 7.5902073
#>  [61,] 1.3070057 6.2832016
#>  [62,] 0.5235666 7.0666408
#>  [63,] 1.3829476 6.2072597
#>  [64,] 0.5235666 7.0666408
#>  [65,] 1.1574218 6.4327855
#>  [66,] 5.3154191 2.2747882
#>  [67,] 5.8370674 1.7531399
#>  [68,] 5.6874834 1.9027239
#>  [69,] 6.0969711 1.4932362
#>  [70,] 6.0969711 1.4932362
#>  [71,] 5.5976724 1.9925349
#>  [72,] 6.2832016 1.3070057
#>  [73,] 5.9354682 1.6547391
#>  [74,] 6.0966517 1.4935556
#>  [75,] 5.3012936 2.2889137
#>  [76,] 5.9741973 1.6160100
#>  [77,] 5.9741973 1.6160100
#>  [78,] 6.0966517 1.4935556
#>  [79,] 6.2459888 1.3442185
#>  [80,] 6.2072597 1.3829476
#>  [81,] 7.5902073 0.0000000
#>  [82,] 6.9563521 0.6338552
#>  [83,] 5.6493464 1.9408609
#>  [84,] 7.5902073 0.0000000
#>  [85,] 6.9563521 0.6338552
#>  [86,] 6.3221775 1.2680298
#>  [87,] 6.2603611 1.3298462
#>  [88,] 6.3421357 1.2480716
#>  [89,] 5.6502707 1.9399367
#>  [90,] 6.2603611 1.3298462
#>  [91,] 6.1697113 1.4204960
#>  [92,] 6.9563521 0.6338552
#>  [93,] 5.4009801 2.1892272
#>  [94,] 6.8942163 0.6959910
#>  [95,] 6.0063213 1.5838860
#>  [96,] 5.6874834 1.9027239
#>  [97,] 5.6874834 1.9027239
#>  [98,] 6.8657023 0.7245050
#>  [99,] 5.4484079 2.1417994
#> [100,] 5.7749316 1.8152757
#> [101,] 7.5902073 0.0000000
#> [102,] 5.4361696 2.1540377
#> [103,] 6.0966517 1.4935556
#> [104,] 6.0450503 1.5451570
#> [105,] 5.4506307 2.1395766
#> [106,] 4.8164561 2.7737512
#> [107,] 5.6374274 1.9527799
#> [108,] 5.4397092 2.1504981
#> [109,] 5.4397092 2.1504981
#> [110,] 5.4397092 2.1504981
#> [111,] 6.0822631 1.5079442
#> [112,] 5.4484079 2.1417994
#> [113,] 5.3724660 2.2177413
#> [114,] 6.1353808 1.4548265
#> [115,] 5.2763191 2.3138882
#> [116,] 7.5902073 0.0000000
#> [117,] 7.0666408 0.5235666
#> [118,] 6.4324661 1.1577412
#> [119,] 4.6815639 2.9086434
#> [120,] 5.6502707 1.9399367
#> [121,] 7.5902073 0.0000000
#> [122,] 6.9563521 0.6338552
#> [123,] 7.5902073 0.0000000
#> [124,] 7.5902073 0.0000000
#> [125,] 7.5902073 0.0000000
#> [126,] 7.5902073 0.0000000
#> [127,] 5.9473560 1.6428513
#> [128,] 5.8370674 1.7531399
#> [129,] 6.9563521 0.6338552
#> [130,] 6.9563521 0.6338552
#> [131,] 6.2603611 1.3298462
#> [132,] 6.4327855 1.1574218
#> [133,] 6.4327855 1.1574218
#> [134,] 6.4327855 1.1574218
#> [135,] 6.4327855 1.1574218
#> [136,] 6.4327855 1.1574218
#> [137,] 6.9563521 0.6338552
#> [138,] 6.9563521 0.6338552
#> [139,] 5.7364752 1.8537321
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.29995314 0.70004686
#>   [2,] 0.34082791 0.65917209
#>   [3,] 0.38087125 0.61912875
#>   [4,] 0.46556898 0.53443102
#>   [5,] 0.30156142 0.69843858
#>   [6,] 0.39515301 0.60484699
#>   [7,] 0.37391729 0.62608271
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.16067513 0.83932487
#>  [10,] 0.24798350 0.75201650
#>  [11,] 0.32122683 0.67877317
#>  [12,] 0.21644353 0.78355647
#>  [13,] 0.41667943 0.58332057
#>  [14,] 0.17709905 0.82290095
#>  [15,] 0.42001596 0.57998404
#>  [16,] 0.21290723 0.78709277
#>  [17,] 0.00000000 1.00000000
#>  [18,] 0.33913583 0.66086417
#>  [19,] 0.09545260 0.90454740
#>  [20,] 0.21290723 0.78709277
#>  [21,] 0.17896221 0.82103779
#>  [22,] 0.37391729 0.62608271
#>  [23,] 0.25570592 0.74429408
#>  [24,] 0.25612773 0.74387227
#>  [25,] 0.39020605 0.60979395
#>  [26,] 0.06897922 0.93102078
#>  [27,] 0.08350961 0.91649039
#>  [28,] 0.21644353 0.78355647
#>  [29,] 0.06897922 0.93102078
#>  [30,] 0.06897922 0.93102078
#>  [31,] 0.31189613 0.68810387
#>  [32,] 0.17709905 0.82290095
#>  [33,] 0.21290723 0.78709277
#>  [34,] 0.30489385 0.69510615
#>  [35,] 0.19162944 0.80837056
#>  [36,] 0.25225171 0.74774829
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.08355169 0.91644831
#>  [39,] 0.08355169 0.91644831
#>  [40,] 0.19167152 0.80832848
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.00000000 1.00000000
#>  [43,] 0.00000000 1.00000000
#>  [44,] 0.16067513 0.83932487
#>  [45,] 0.09169591 0.90830409
#>  [46,] 0.06897922 0.93102078
#>  [47,] 0.00000000 1.00000000
#>  [48,] 0.10811983 0.89188017
#>  [49,] 0.15253090 0.84746910
#>  [50,] 0.00000000 1.00000000
#>  [51,] 0.11322234 0.88677766
#>  [52,] 0.17524760 0.82475240
#>  [53,] 0.19491301 0.80508699
#>  [54,] 0.34613401 0.65386599
#>  [55,] 0.34237732 0.65762268
#>  [56,] 0.39325734 0.60674266
#>  [57,] 0.47870469 0.52129531
#>  [58,] 0.28003500 0.71996500
#>  [59,] 0.29488861 0.70511139
#>  [60,] 0.00000000 1.00000000
#>  [61,] 0.17219631 0.82780369
#>  [62,] 0.06897922 0.93102078
#>  [63,] 0.18220155 0.81779845
#>  [64,] 0.06897922 0.93102078
#>  [65,] 0.15248882 0.84751118
#>  [66,] 0.70029959 0.29970041
#>  [67,] 0.76902608 0.23097392
#>  [68,] 0.74931859 0.25068141
#>  [69,] 0.80326806 0.19673194
#>  [70,] 0.80326806 0.19673194
#>  [71,] 0.73748611 0.26251389
#>  [72,] 0.82780369 0.17219631
#>  [73,] 0.78199026 0.21800974
#>  [74,] 0.80322598 0.19677402
#>  [75,] 0.69843858 0.30156142
#>  [76,] 0.78709277 0.21290723
#>  [77,] 0.78709277 0.21290723
#>  [78,] 0.80322598 0.19677402
#>  [79,] 0.82290095 0.17709905
#>  [80,] 0.81779845 0.18220155
#>  [81,] 1.00000000 0.00000000
#>  [82,] 0.91649039 0.08350961
#>  [83,] 0.74429408 0.25570592
#>  [84,] 1.00000000 0.00000000
#>  [85,] 0.91649039 0.08350961
#>  [86,] 0.83293871 0.16706129
#>  [87,] 0.82479448 0.17520552
#>  [88,] 0.83556818 0.16443182
#>  [89,] 0.74441585 0.25558415
#>  [90,] 0.82479448 0.17520552
#>  [91,] 0.81285149 0.18714851
#>  [92,] 0.91649039 0.08350961
#>  [93,] 0.71157214 0.28842786
#>  [94,] 0.90830409 0.09169591
#>  [95,] 0.79132506 0.20867494
#>  [96,] 0.74931859 0.25068141
#>  [97,] 0.74931859 0.25068141
#>  [98,] 0.90454740 0.09545260
#>  [99,] 0.71782070 0.28217930
#> [100,] 0.76083977 0.23916023
#> [101,] 1.00000000 0.00000000
#> [102,] 0.71620831 0.28379169
#> [103,] 0.80322598 0.19677402
#> [104,] 0.79642757 0.20357243
#> [105,] 0.71811355 0.28188645
#> [106,] 0.63456187 0.36543813
#> [107,] 0.74272378 0.25727622
#> [108,] 0.71667465 0.28332535
#> [109,] 0.71667465 0.28332535
#> [110,] 0.71667465 0.28332535
#> [111,] 0.80133030 0.19866970
#> [112,] 0.71782070 0.28217930
#> [113,] 0.70781546 0.29218454
#> [114,] 0.80832848 0.19167152
#> [115,] 0.69514822 0.30485178
#> [116,] 1.00000000 0.00000000
#> [117,] 0.93102078 0.06897922
#> [118,] 0.84746910 0.15253090
#> [119,] 0.61678999 0.38321001
#> [120,] 0.74441585 0.25558415
#> [121,] 1.00000000 0.00000000
#> [122,] 0.91649039 0.08350961
#> [123,] 1.00000000 0.00000000
#> [124,] 1.00000000 0.00000000
#> [125,] 1.00000000 0.00000000
#> [126,] 1.00000000 0.00000000
#> [127,] 0.78355647 0.21644353
#> [128,] 0.76902608 0.23097392
#> [129,] 0.91649039 0.08350961
#> [130,] 0.91649039 0.08350961
#> [131,] 0.82479448 0.17520552
#> [132,] 0.84751118 0.15248882
#> [133,] 0.84751118 0.15248882
#> [134,] 0.84751118 0.15248882
#> [135,] 0.84751118 0.15248882
#> [136,] 0.84751118 0.15248882
#> [137,] 0.91649039 0.08350961
#> [138,] 0.91649039 0.08350961
#> [139,] 0.75577319 0.24422681
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  0.0000000 14.5635108  3.6070387  0.0000000  1.5322495  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  1.0605317  0.0000000  0.0000000  4.3510818  2.7524261  6.2907381 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  0.9233629  0.0000000  0.0000000  1.3421698  0.0000000  7.2227656 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  3.0335548  0.0000000  1.6832527  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  4.4102744  5.0107010  6.5254015  0.0000000  1.0897319  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  5.4055095  0.0000000  2.1892109  2.6336620  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.2188329  2.4431273  2.4489635  0.0000000  4.0912994  1.9027738  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  1.2482368  0.0000000  0.0000000  0.0000000  0.0000000  1.8836603  5.1153032 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  2.0206293  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 74 65 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V28        V37        V21        V43         V6        V36 
#> 14.5635108  7.2227656  6.5254015  6.2907381  5.4055095  5.1153032  5.0107010 
#>        V35         V2        V51        V12        V48        V31        V20 
#>  4.4102744  4.3510818  4.0912994  3.6070387  3.2188329  3.0335548  2.7524261 
#>        V46         V5        V49        V45         V8        V52        V59 
#>  2.6336620  2.4489635  2.4431273  2.1892109  2.0206293  1.9027738  1.8836603 
#>        V33        V14        V26        V54        V39        V17        V23 
#>  1.6832527  1.5322495  1.3421698  1.2482368  1.0897319  1.0605317  0.9233629 
#>         V1        V10        V13        V15        V16        V18        V19 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V22        V24        V25        V27        V29         V3        V30 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V32        V34        V38         V4        V40        V41        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V44        V47        V50        V53        V55        V56        V57 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V58        V60         V7         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1449275