Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 . https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V11>=0.19645 79 20 M (0.74683544 0.25316456)  
#>      4) V18< 0.4129 46  1 M (0.97826087 0.02173913) *
#>      5) V18>=0.4129 33 14 R (0.42424242 0.57575758)  
#>       10) V37< 0.3766 16  3 M (0.81250000 0.18750000) *
#>       11) V37>=0.3766 17  1 R (0.05882353 0.94117647) *
#>    3) V11< 0.19645 60  7 R (0.11666667 0.88333333)  
#>      6) V7>=0.14045 8  2 M (0.75000000 0.25000000) *
#>      7) V7< 0.14045 52  1 R (0.01923077 0.98076923) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V21>=0.6827 78 22 M (0.71794872 0.28205128)  
#>      4) V44>=0.12675 58  8 M (0.86206897 0.13793103)  
#>        8) V31< 0.55405 51  2 M (0.96078431 0.03921569) *
#>        9) V31>=0.55405 7  1 R (0.14285714 0.85714286) *
#>      5) V44< 0.12675 20  6 R (0.30000000 0.70000000)  
#>       10) V12>=0.34615 7  2 M (0.71428571 0.28571429) *
#>       11) V12< 0.34615 13  1 R (0.07692308 0.92307692) *
#>    3) V21< 0.6827 61 12 R (0.19672131 0.80327869)  
#>      6) V12>=0.2438 21  9 M (0.57142857 0.42857143)  
#>       12) V48>=0.07645 11  0 M (1.00000000 0.00000000) *
#>       13) V48< 0.07645 10  1 R (0.10000000 0.90000000) *
#>      7) V12< 0.2438 40  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 M (0.51079137 0.48920863)  
#>    2) V52>=0.0069 106 37 M (0.65094340 0.34905660)  
#>      4) V17< 0.6968 78 14 M (0.82051282 0.17948718)  
#>        8) V51>=0.008 66  6 M (0.90909091 0.09090909) *
#>        9) V51< 0.008 12  4 R (0.33333333 0.66666667) *
#>      5) V17>=0.6968 28  5 R (0.17857143 0.82142857)  
#>       10) V19>=0.86695 7  2 M (0.71428571 0.28571429) *
#>       11) V19< 0.86695 21  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0069 33  2 R (0.06060606 0.93939394) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 R (0.48201439 0.51798561)  
#>    2) V45>=0.19565 44  4 M (0.90909091 0.09090909) *
#>    3) V45< 0.19565 95 27 R (0.28421053 0.71578947)  
#>      6) V11>=0.2481 35 14 M (0.60000000 0.40000000)  
#>       12) V21< 0.81025 18  1 M (0.94444444 0.05555556) *
#>       13) V21>=0.81025 17  4 R (0.23529412 0.76470588) *
#>      7) V11< 0.2481 60  6 R (0.10000000 0.90000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V46>=0.06905 93 33 M (0.64516129 0.35483871)  
#>      4) V39>=0.21875 64 11 M (0.82812500 0.17187500)  
#>        8) V52>=0.00495 57  4 M (0.92982456 0.07017544) *
#>        9) V52< 0.00495 7  0 R (0.00000000 1.00000000) *
#>      5) V39< 0.21875 29  7 R (0.24137931 0.75862069)  
#>       10) V37< 0.18075 8  2 M (0.75000000 0.25000000) *
#>       11) V37>=0.18075 21  1 R (0.04761905 0.95238095) *
#>    3) V46< 0.06905 46  3 R (0.06521739 0.93478261) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 63 R (0.45323741 0.54676259)  
#>   2) V11>=0.19795 80 27 M (0.66250000 0.33750000)  
#>     4) V17< 0.92805 70 17 M (0.75714286 0.24285714)  
#>       8) V31>=0.25015 60  9 M (0.85000000 0.15000000) *
#>       9) V31< 0.25015 10  2 R (0.20000000 0.80000000) *
#>     5) V17>=0.92805 10  0 R (0.00000000 1.00000000) *
#>   3) V11< 0.19795 59 10 R (0.16949153 0.83050847)  
#>     6) V33< 0.21985 8  2 M (0.75000000 0.25000000) *
#>     7) V33>=0.21985 51  4 R (0.07843137 0.92156863) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V49>=0.02395 101 39 M (0.61386139 0.38613861)  
#>      4) V3>=0.02025 72 15 M (0.79166667 0.20833333)  
#>        8) V12>=0.12395 65  8 M (0.87692308 0.12307692)  
#>         16) V52>=0.007 56  1 M (0.98214286 0.01785714) *
#>         17) V52< 0.007 9  2 R (0.22222222 0.77777778) *
#>        9) V12< 0.12395 7  0 R (0.00000000 1.00000000) *
#>      5) V3< 0.02025 29  5 R (0.17241379 0.82758621)  
#>       10) V47>=0.1304 7  2 M (0.71428571 0.28571429) *
#>       11) V47< 0.1304 22  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.02395 38  1 R (0.02631579 0.97368421) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V23>=0.80595 52  9 M (0.82692308 0.17307692)  
#>      4) V13>=0.16055 42  2 M (0.95238095 0.04761905) *
#>      5) V13< 0.16055 10  3 R (0.30000000 0.70000000) *
#>    3) V23< 0.80595 87 23 R (0.26436782 0.73563218)  
#>      6) V55< 0.0045 15  3 M (0.80000000 0.20000000) *
#>      7) V55>=0.0045 72 11 R (0.15277778 0.84722222)  
#>       14) V51>=0.024 9  2 M (0.77777778 0.22222222) *
#>       15) V51< 0.024 63  4 R (0.06349206 0.93650794)  
#>         30) V44< 0.0745 7  3 M (0.57142857 0.42857143) *
#>         31) V44>=0.0745 56  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 R (0.44604317 0.55395683)  
#>    2) V55< 0.005 42  8 M (0.80952381 0.19047619)  
#>      4) V20>=0.39705 33  2 M (0.93939394 0.06060606) *
#>      5) V20< 0.39705 9  3 R (0.33333333 0.66666667) *
#>    3) V55>=0.005 97 28 R (0.28865979 0.71134021)  
#>      6) V45>=0.2726 12  0 M (1.00000000 0.00000000) *
#>      7) V45< 0.2726 85 16 R (0.18823529 0.81176471)  
#>       14) V23>=0.8354 9  1 M (0.88888889 0.11111111) *
#>       15) V23< 0.8354 76  8 R (0.10526316 0.89473684)  
#>         30) V28>=0.94295 9  4 M (0.55555556 0.44444444) *
#>         31) V28< 0.94295 67  3 R (0.04477612 0.95522388) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.5323741 0.4676259)  
#>    2) V59>=0.0123 32  3 M (0.9062500 0.0937500) *
#>    3) V59< 0.0123 107 45 R (0.4205607 0.5794393)  
#>      6) V55< 0.0063 50 15 M (0.7000000 0.3000000)  
#>       12) V11>=0.1019 40  5 M (0.8750000 0.1250000) *
#>       13) V11< 0.1019 10  0 R (0.0000000 1.0000000) *
#>      7) V55>=0.0063 57 10 R (0.1754386 0.8245614)  
#>       14) V23>=0.8354 17  8 M (0.5294118 0.4705882) *
#>       15) V23< 0.8354 40  1 R (0.0250000 0.9750000) *
#> 
#> 
#> $weights
#>  [1] 0.8630811 0.9330391 0.7532239 0.7246756 0.8080347 0.7005606 0.7876514
#>  [8] 0.8944247 0.8275013 0.7557008
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 1.5877567 6.4601365
#>   [2,] 1.4252363 6.6226570
#>   [3,] 2.2527376 5.7951557
#>   [4,] 3.2023024 4.8455908
#>   [5,] 2.3364260 5.7114673
#>   [6,] 1.5433522 6.5045410
#>   [7,] 2.3364260 5.7114673
#>   [8,] 0.0000000 8.0478933
#>   [9,] 2.4556833 5.5922100
#>  [10,] 1.5637355 6.4841577
#>  [11,] 2.3513870 5.6965063
#>  [12,] 1.4562614 6.5916318
#>  [13,] 0.8080347 7.2398585
#>  [14,] 2.5837716 5.4641217
#>  [15,] 2.3054008 5.7424925
#>  [16,] 0.0000000 8.0478933
#>  [17,] 1.6191003 6.4287930
#>  [18,] 2.4580664 5.5898269
#>  [19,] 2.3489100 5.6989832
#>  [20,] 2.2618192 5.7860740
#>  [21,] 0.0000000 8.0478933
#>  [22,] 1.6711158 6.3767774
#>  [23,] 2.4353000 5.6125933
#>  [24,] 3.1995071 4.8483862
#>  [25,] 2.3364260 5.7114673
#>  [26,] 1.5832021 6.4646912
#>  [27,] 2.4033493 5.6445439
#>  [28,] 0.0000000 8.0478933
#>  [29,] 1.5327104 6.5151829
#>  [30,] 2.4030201 5.6448732
#>  [31,] 1.6335997 6.4142936
#>  [32,] 2.4942977 5.5535956
#>  [33,] 2.6151152 5.4327781
#>  [34,] 0.7005606 7.3473326
#>  [35,] 3.1134356 4.9344577
#>  [36,] 0.0000000 8.0478933
#>  [37,] 0.7005606 7.3473326
#>  [38,] 0.0000000 8.0478933
#>  [39,] 2.4776268 5.5702665
#>  [40,] 1.6355361 6.4123572
#>  [41,] 2.4776268 5.5702665
#>  [42,] 0.8080347 7.2398585
#>  [43,] 1.5832021 6.4646912
#>  [44,] 0.0000000 8.0478933
#>  [45,] 2.3364260 5.7114673
#>  [46,] 0.8275013 7.2203920
#>  [47,] 1.7219260 6.3259673
#>  [48,] 0.0000000 8.0478933
#>  [49,] 1.5956862 6.4522071
#>  [50,] 2.3748011 5.6730922
#>  [51,] 1.4537845 6.5941088
#>  [52,] 1.5085954 6.5392979
#>  [53,] 2.4030201 5.6448732
#>  [54,] 1.5280620 6.5198313
#>  [55,] 0.8275013 7.2203920
#>  [56,] 2.5137642 5.5341291
#>  [57,] 1.7961202 6.2517731
#>  [58,] 2.4776268 5.5702665
#>  [59,] 1.5636417 6.4842515
#>  [60,] 2.5837716 5.4641217
#>  [61,] 1.5877567 6.4601365
#>  [62,] 0.8630811 7.1848122
#>  [63,] 1.7575058 6.2903875
#>  [64,] 2.3196610 5.7282323
#>  [65,] 2.1809371 5.8669562
#>  [66,] 0.7246756 7.3232176
#>  [67,] 1.5612586 6.4866347
#>  [68,] 2.2883174 5.7595759
#>  [69,] 1.6163050 6.4315883
#>  [70,] 0.0000000 8.0478933
#>  [71,] 6.2204295 1.8274637
#>  [72,] 6.3259673 1.7219260
#>  [73,] 6.4123572 1.6355361
#>  [74,] 4.8690050 3.1788883
#>  [75,] 4.9000301 3.1478631
#>  [76,] 5.7595759 2.2883174
#>  [77,] 6.5941088 1.4537845
#>  [78,] 5.6411415 2.4067518
#>  [79,] 7.1848122 0.8630811
#>  [80,] 6.4601365 1.5877567
#>  [81,] 6.4601365 1.5877567
#>  [82,] 6.6226570 1.4252363
#>  [83,] 7.3232176 0.7246756
#>  [84,] 6.6226570 1.4252363
#>  [85,] 7.3232176 0.7246756
#>  [86,] 6.5151829 1.5327104
#>  [87,] 7.2602418 0.7876514
#>  [88,] 6.5151829 1.5327104
#>  [89,] 6.5699938 1.4778995
#>  [90,] 6.5198313 1.5280620
#>  [91,] 5.6101164 2.4377769
#>  [92,] 6.4529080 1.5949853
#>  [93,] 6.3977678 1.6501255
#>  [94,] 5.7321799 2.3157134
#>  [95,] 6.2204295 1.8274637
#>  [96,] 7.2398585 0.8080347
#>  [97,] 7.2921925 0.7557008
#>  [98,] 7.1148542 0.9330391
#>  [99,] 8.0478933 0.0000000
#> [100,] 7.1148542 0.9330391
#> [101,] 7.3232176 0.7246756
#> [102,] 5.6012916 2.4466016
#> [103,] 5.8350055 2.2128877
#> [104,] 5.8064573 2.2414360
#> [105,] 4.9095557 3.1383376
#> [106,] 7.2203920 0.8275013
#> [107,] 7.2921925 0.7557008
#> [108,] 6.4671681 1.5807252
#> [109,] 7.2921925 0.7557008
#> [110,] 6.3272028 1.7206905
#> [111,] 6.2517731 1.7961202
#> [112,] 5.7310277 2.3168656
#> [113,] 6.3616304 1.6862629
#> [114,] 6.4601365 1.5877567
#> [115,] 5.6266421 2.4212512
#> [116,] 5.6989832 2.3489100
#> [117,] 6.5070180 1.5408753
#> [118,] 6.5070180 1.5408753
#> [119,] 5.7595759 2.2883174
#> [120,] 7.2398585 0.8080347
#> [121,] 7.2921925 0.7557008
#> [122,] 7.2921925 0.7557008
#> [123,] 7.2921925 0.7557008
#> [124,] 6.3591534 1.6887399
#> [125,] 7.2921925 0.7557008
#> [126,] 7.2921925 0.7557008
#> [127,] 7.2946694 0.7532239
#> [128,] 6.3454339 1.7024594
#> [129,] 7.2203920 0.8275013
#> [130,] 8.0478933 0.0000000
#> [131,] 6.5675168 1.4803764
#> [132,] 6.3977678 1.6501255
#> [133,] 8.0478933 0.0000000
#> [134,] 8.0478933 0.0000000
#> [135,] 8.0478933 0.0000000
#> [136,] 6.4123572 1.6355361
#> [137,] 7.2203920 0.8275013
#> [138,] 7.2946694 0.7532239
#> [139,] 5.5191680 2.5287252
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.19728849 0.80271151
#>   [2,] 0.17709433 0.82290567
#>   [3,] 0.27991644 0.72008356
#>   [4,] 0.39790568 0.60209432
#>   [5,] 0.29031523 0.70968477
#>   [6,] 0.19177096 0.80822904
#>   [7,] 0.29031523 0.70968477
#>   [8,] 0.00000000 1.00000000
#>   [9,] 0.30513368 0.69486632
#>  [10,] 0.19430371 0.80569629
#>  [11,] 0.29217423 0.70782577
#>  [12,] 0.18094940 0.81905060
#>  [13,] 0.10040326 0.89959674
#>  [14,] 0.32104944 0.67895056
#>  [15,] 0.28646016 0.71353984
#>  [16,] 0.00000000 1.00000000
#>  [17,] 0.20118313 0.79881687
#>  [18,] 0.30542980 0.69457020
#>  [19,] 0.29186645 0.70813355
#>  [20,] 0.28104489 0.71895511
#>  [21,] 0.00000000 1.00000000
#>  [22,] 0.20764637 0.79235363
#>  [23,] 0.30260093 0.69739907
#>  [24,] 0.39755834 0.60244166
#>  [25,] 0.29031523 0.70968477
#>  [26,] 0.19672255 0.80327745
#>  [27,] 0.29863087 0.70136913
#>  [28,] 0.00000000 1.00000000
#>  [29,] 0.19044865 0.80955135
#>  [30,] 0.29858995 0.70141005
#>  [31,] 0.20298476 0.79701524
#>  [32,] 0.30993175 0.69006825
#>  [33,] 0.32494407 0.67505593
#>  [34,] 0.08704895 0.91295105
#>  [35,] 0.38686343 0.61313657
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.08704895 0.91295105
#>  [38,] 0.00000000 1.00000000
#>  [39,] 0.30786029 0.69213971
#>  [40,] 0.20322537 0.79677463
#>  [41,] 0.30786029 0.69213971
#>  [42,] 0.10040326 0.89959674
#>  [43,] 0.19672255 0.80327745
#>  [44,] 0.00000000 1.00000000
#>  [45,] 0.29031523 0.70968477
#>  [46,] 0.10282210 0.89717790
#>  [47,] 0.21395984 0.78604016
#>  [48,] 0.00000000 1.00000000
#>  [49,] 0.19827378 0.80172622
#>  [50,] 0.29508358 0.70491642
#>  [51,] 0.18064162 0.81935838
#>  [52,] 0.18745221 0.81254779
#>  [53,] 0.29858995 0.70141005
#>  [54,] 0.18987105 0.81012895
#>  [55,] 0.10282210 0.89717790
#>  [56,] 0.31235059 0.68764941
#>  [57,] 0.22317892 0.77682108
#>  [58,] 0.30786029 0.69213971
#>  [59,] 0.19429206 0.80570794
#>  [60,] 0.32104944 0.67895056
#>  [61,] 0.19728849 0.80271151
#>  [62,] 0.10724311 0.89275689
#>  [63,] 0.21838085 0.78161915
#>  [64,] 0.28823207 0.71176793
#>  [65,] 0.27099478 0.72900522
#>  [66,] 0.09004539 0.90995461
#>  [67,] 0.19399594 0.80600406
#>  [68,] 0.28433744 0.71566256
#>  [69,] 0.20083578 0.79916422
#>  [70,] 0.00000000 1.00000000
#>  [71,] 0.77292645 0.22707355
#>  [72,] 0.78604016 0.21395984
#>  [73,] 0.79677463 0.20322537
#>  [74,] 0.60500367 0.39499633
#>  [75,] 0.60885874 0.39114126
#>  [76,] 0.71566256 0.28433744
#>  [77,] 0.81935838 0.18064162
#>  [78,] 0.70094636 0.29905364
#>  [79,] 0.89275689 0.10724311
#>  [80,] 0.80271151 0.19728849
#>  [81,] 0.80271151 0.19728849
#>  [82,] 0.82290567 0.17709433
#>  [83,] 0.90995461 0.09004539
#>  [84,] 0.82290567 0.17709433
#>  [85,] 0.90995461 0.09004539
#>  [86,] 0.80955135 0.19044865
#>  [87,] 0.90212949 0.09787051
#>  [88,] 0.80955135 0.19044865
#>  [89,] 0.81636194 0.18363806
#>  [90,] 0.81012895 0.18987105
#>  [91,] 0.69709130 0.30290870
#>  [92,] 0.80181331 0.19818669
#>  [93,] 0.79496181 0.20503819
#>  [94,] 0.71225844 0.28774156
#>  [95,] 0.77292645 0.22707355
#>  [96,] 0.89959674 0.10040326
#>  [97,] 0.90609955 0.09390045
#>  [98,] 0.88406419 0.11593581
#>  [99,] 1.00000000 0.00000000
#> [100,] 0.88406419 0.11593581
#> [101,] 0.90995461 0.09004539
#> [102,] 0.69599477 0.30400523
#> [103,] 0.72503515 0.27496485
#> [104,] 0.72148786 0.27851214
#> [105,] 0.61004235 0.38995765
#> [106,] 0.89717790 0.10282210
#> [107,] 0.90609955 0.09390045
#> [108,] 0.80358522 0.19641478
#> [109,] 0.90609955 0.09390045
#> [110,] 0.78619367 0.21380633
#> [111,] 0.77682108 0.22317892
#> [112,] 0.71211527 0.28788473
#> [113,] 0.79047151 0.20952849
#> [114,] 0.80271151 0.19728849
#> [115,] 0.69914472 0.30085528
#> [116,] 0.70813355 0.29186645
#> [117,] 0.80853681 0.19146319
#> [118,] 0.80853681 0.19146319
#> [119,] 0.71566256 0.28433744
#> [120,] 0.89959674 0.10040326
#> [121,] 0.90609955 0.09390045
#> [122,] 0.90609955 0.09390045
#> [123,] 0.90609955 0.09390045
#> [124,] 0.79016374 0.20983626
#> [125,] 0.90609955 0.09390045
#> [126,] 0.90609955 0.09390045
#> [127,] 0.90640732 0.09359268
#> [128,] 0.78845900 0.21154100
#> [129,] 0.89717790 0.10282210
#> [130,] 1.00000000 0.00000000
#> [131,] 0.81605417 0.18394583
#> [132,] 0.79496181 0.20503819
#> [133,] 1.00000000 0.00000000
#> [134,] 1.00000000 0.00000000
#> [135,] 1.00000000 0.00000000
#> [136,] 0.79677463 0.20322537
#> [137,] 0.89717790 0.10282210
#> [138,] 0.90640732 0.09359268
#> [139,] 0.68579041 0.31420959
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.0000000  0.0000000 12.9636728  4.8409987  1.5269563  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  0.0000000  4.9213073  2.5280117  1.0019860  0.0000000  1.0674371  5.8915862 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  7.7451911  0.0000000  0.0000000  0.0000000  0.0000000  0.8506944 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  3.1010845  0.0000000  3.1680805  0.0000000  1.0850982  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  0.0000000  3.1537194  0.0000000  2.7571425  0.0000000  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  0.0000000  3.0795926  7.0737209  4.1539602  1.0598272 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  1.9660410  3.7285652  0.0000000  0.0000000  3.0438538  7.1952425  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  8.3283658  0.0000000  0.0000000  0.0000000  2.1808217  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  1.5870424  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 69 70 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11        V55        V23        V52        V45        V21        V17 
#> 12.9636728  8.3283658  7.7451911  7.1952425  7.0737209  5.8915862  4.9213073 
#>        V12        V46        V49        V31        V37         V3        V44 
#>  4.8409987  4.1539602  3.7285652  3.1680805  3.1537194  3.1010845  3.0795926 
#>        V51        V39        V18        V59        V48         V7        V13 
#>  3.0438538  2.7571425  2.5280117  2.1808217  1.9660410  1.5870424  1.5269563 
#>        V33        V20        V47        V19        V28         V1        V10 
#>  1.0850982  1.0674371  1.0598272  1.0019860  0.8506944  0.0000000  0.0000000 
#>        V14        V15        V16         V2        V22        V24        V25 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V26        V27        V29        V30        V32        V34        V35 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V36        V38         V4        V40        V41        V42        V43 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V5        V50        V53        V54        V56        V57        V58 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>         V6        V60         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696