Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifAdabag$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

LearnerClassifAdabag$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 M (0.5323741 0.4676259)  
#>    2) V9>=0.10785 103 33 M (0.6796117 0.3203883)  
#>      4) V43>=0.19835 59  7 M (0.8813559 0.1186441)  
#>        8) V37< 0.5718 46  0 M (1.0000000 0.0000000) *
#>        9) V37>=0.5718 13  6 R (0.4615385 0.5384615) *
#>      5) V43< 0.19835 44 18 R (0.4090909 0.5909091)  
#>       10) V23>=0.7953 17  3 M (0.8235294 0.1764706) *
#>       11) V23< 0.7953 27  4 R (0.1481481 0.8518519)  
#>         22) V7< 0.08965 7  3 M (0.5714286 0.4285714) *
#>         23) V7>=0.08965 20  0 R (0.0000000 1.0000000) *
#>    3) V9< 0.10785 36  4 R (0.1111111 0.8888889)  
#>      6) V19>=0.74895 7  3 M (0.5714286 0.4285714) *
#>      7) V19< 0.74895 29  0 R (0.0000000 1.0000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V49>=0.0499 62 14 M (0.77419355 0.22580645)  
#>      4) V9>=0.1002 55  7 M (0.87272727 0.12727273)  
#>        8) V15< 0.6077 47  2 M (0.95744681 0.04255319) *
#>        9) V15>=0.6077 8  3 R (0.37500000 0.62500000) *
#>      5) V9< 0.1002 7  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.0499 77 15 R (0.19480519 0.80519481)  
#>      6) V20>=0.49225 40 15 R (0.37500000 0.62500000)  
#>       12) V46>=0.07745 20  7 M (0.65000000 0.35000000)  
#>         24) V55< 0.0038 9  0 M (1.00000000 0.00000000) *
#>         25) V55>=0.0038 11  4 R (0.36363636 0.63636364) *
#>       13) V46< 0.07745 20  2 R (0.10000000 0.90000000) *
#>      7) V20< 0.49225 37  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 67 M (0.51798561 0.48201439)  
#>    2) V12>=0.20075 87 24 M (0.72413793 0.27586207)  
#>      4) V9>=0.10115 69 11 M (0.84057971 0.15942029)  
#>        8) V5< 0.1526 62  6 M (0.90322581 0.09677419)  
#>         16) V22>=0.3009 53  1 M (0.98113208 0.01886792) *
#>         17) V22< 0.3009 9  4 R (0.44444444 0.55555556) *
#>        9) V5>=0.1526 7  2 R (0.28571429 0.71428571) *
#>      5) V9< 0.10115 18  5 R (0.27777778 0.72222222) *
#>    3) V12< 0.20075 52  9 R (0.17307692 0.82692308)  
#>      6) V21>=0.70685 12  4 M (0.66666667 0.33333333) *
#>      7) V21< 0.70685 40  1 R (0.02500000 0.97500000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 53 M (0.61870504 0.38129496)  
#>    2) V11>=0.11445 116 31 M (0.73275862 0.26724138)  
#>      4) V3< 0.0389 51  3 M (0.94117647 0.05882353) *
#>      5) V3>=0.0389 65 28 M (0.56923077 0.43076923)  
#>       10) V55>=0.0124 32  3 M (0.90625000 0.09375000) *
#>       11) V55< 0.0124 33  8 R (0.24242424 0.75757576)  
#>         22) V54>=0.01325 8  1 M (0.87500000 0.12500000) *
#>         23) V54< 0.01325 25  1 R (0.04000000 0.96000000) *
#>    3) V11< 0.11445 23  1 R (0.04347826 0.95652174) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 56 M (0.59712230 0.40287770)  
#>   2) V52>=0.00905 92 20 M (0.78260870 0.21739130)  
#>     4) V49>=0.02385 77  9 M (0.88311688 0.11688312)  
#>       8) V11>=0.1261 70  3 M (0.95714286 0.04285714) *
#>       9) V11< 0.1261 7  1 R (0.14285714 0.85714286) *
#>     5) V49< 0.02385 15  4 R (0.26666667 0.73333333) *
#>   3) V52< 0.00905 47 11 R (0.23404255 0.76595745)  
#>     6) V23>=0.7505 13  3 M (0.76923077 0.23076923) *
#>     7) V23< 0.7505 34  1 R (0.02941176 0.97058824) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V52>=0.0088 91 31 M (0.65934066 0.34065934)  
#>      4) V36< 0.31835 46  4 M (0.91304348 0.08695652) *
#>      5) V36>=0.31835 45 18 R (0.40000000 0.60000000)  
#>       10) V45>=0.2711 14  1 M (0.92857143 0.07142857) *
#>       11) V45< 0.2711 31  5 R (0.16129032 0.83870968)  
#>         22) V19>=0.6797 7  2 M (0.71428571 0.28571429) *
#>         23) V19< 0.6797 24  0 R (0.00000000 1.00000000) *
#>    3) V52< 0.0088 48  8 R (0.16666667 0.83333333)  
#>      6) V21>=0.79525 7  2 M (0.71428571 0.28571429) *
#>      7) V21< 0.79525 41  3 R (0.07317073 0.92682927) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 69 R (0.49640288 0.50359712)  
#>    2) V11>=0.1714 93 32 M (0.65591398 0.34408602)  
#>      4) V18< 0.5362 43  2 M (0.95348837 0.04651163) *
#>      5) V18>=0.5362 50 20 R (0.40000000 0.60000000)  
#>       10) V35< 0.28785 26  6 M (0.76923077 0.23076923)  
#>         20) V53>=0.0094 18  0 M (1.00000000 0.00000000) *
#>         21) V53< 0.0094 8  2 R (0.25000000 0.75000000) *
#>       11) V35>=0.28785 24  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1714 46  8 R (0.17391304 0.82608696)  
#>      6) V4>=0.0515 9  1 M (0.88888889 0.11111111) *
#>      7) V4< 0.0515 37  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 65 R (0.46762590 0.53237410)  
#>    2) V2>=0.0677 24  1 M (0.95833333 0.04166667) *
#>    3) V2< 0.0677 115 42 R (0.36521739 0.63478261)  
#>      6) V35< 0.19285 31  9 M (0.70967742 0.29032258)  
#>       12) V10>=0.1812 19  1 M (0.94736842 0.05263158) *
#>       13) V10< 0.1812 12  4 R (0.33333333 0.66666667) *
#>      7) V35>=0.19285 84 20 R (0.23809524 0.76190476)  
#>       14) V8< 0.0532 7  1 M (0.85714286 0.14285714) *
#>       15) V8>=0.0532 77 14 R (0.18181818 0.81818182)  
#>         30) V4>=0.06185 14  6 M (0.57142857 0.42857143) *
#>         31) V4< 0.06185 63  6 R (0.09523810 0.90476190)  
#>           62) V31>=0.7744 7  3 M (0.57142857 0.42857143) *
#>           63) V31< 0.7744 56  2 R (0.03571429 0.96428571) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V12>=0.19455 90 32 M (0.64444444 0.35555556)  
#>      4) V7< 0.2124 77 19 M (0.75324675 0.24675325)  
#>        8) V33< 0.3662 46  2 M (0.95652174 0.04347826) *
#>        9) V33>=0.3662 31 14 R (0.45161290 0.54838710)  
#>         18) V31>=0.76115 9  0 M (1.00000000 0.00000000) *
#>         19) V31< 0.76115 22  5 R (0.22727273 0.77272727)  
#>           38) V3< 0.01845 7  2 M (0.71428571 0.28571429) *
#>           39) V3>=0.01845 15  0 R (0.00000000 1.00000000) *
#>      5) V7>=0.2124 13  0 R (0.00000000 1.00000000) *
#>    3) V12< 0.19455 49  5 R (0.10204082 0.89795918) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 R (0.42446043 0.57553957)  
#>    2) V52>=0.024 25  3 M (0.88000000 0.12000000) *
#>    3) V52< 0.024 114 37 R (0.32456140 0.67543860)  
#>      6) V10>=0.18455 50 23 M (0.54000000 0.46000000)  
#>       12) V7< 0.2092 37 10 M (0.72972973 0.27027027)  
#>         24) V52>=0.00885 22  0 M (1.00000000 0.00000000) *
#>         25) V52< 0.00885 15  5 R (0.33333333 0.66666667) *
#>       13) V7>=0.2092 13  0 R (0.00000000 1.00000000) *
#>      7) V10< 0.18455 64 10 R (0.15625000 0.84375000)  
#>       14) V56>=0.0111 9  3 M (0.66666667 0.33333333) *
#>       15) V56< 0.0111 55  4 R (0.07272727 0.92727273)  
#>         30) V16< 0.1524 7  3 M (0.57142857 0.42857143) *
#>         31) V16>=0.1524 48  0 R (0.00000000 1.00000000) *
#> 
#> 
#> $weights
#>  [1] 0.6886628 0.8008090 0.7472668 0.6242054 0.7349164 0.7161467 0.9645618
#>  [8] 0.7131340 0.7157781 0.5771127
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 0.6242054 6.6583883
#>   [2,] 1.4604008 5.8221929
#>   [3,] 2.8007528 4.4818408
#>   [4,] 0.7161467 6.5664469
#>   [5,] 1.4630449 5.8195488
#>   [6,] 2.4276067 4.8549869
#>   [7,] 1.2902467 5.9923469
#>   [8,] 1.4634135 5.8191802
#>   [9,] 1.5887672 5.6938265
#>  [10,] 1.3373393 5.9452543
#>  [11,] 2.3019011 4.9806925
#>  [12,] 2.9131467 4.3694469
#>  [13,] 1.3373393 5.9452543
#>  [14,] 2.1641970 5.1183966
#>  [15,] 2.1367132 5.1458805
#>  [16,] 2.9646374 4.3179563
#>  [17,] 2.0752684 5.2073252
#>  [18,] 0.7131340 6.5694597
#>  [19,] 3.3681593 3.9144344
#>  [20,] 2.8003842 4.4822094
#>  [21,] 2.3045453 4.9780484
#>  [22,] 2.3019011 4.9806925
#>  [23,] 0.7472668 6.5353269
#>  [24,] 3.0851495 4.1974442
#>  [25,] 2.1381483 5.1444454
#>  [26,] 2.2303374 5.0522562
#>  [27,] 1.2013181 6.0812755
#>  [28,] 1.3591217 5.9234719
#>  [29,] 1.3373393 5.9452543
#>  [30,] 1.3591217 5.9234719
#>  [31,] 1.9789096 5.3036841
#>  [32,] 0.7131340 6.5694597
#>  [33,] 0.6242054 6.6583883
#>  [34,] 2.0752684 5.2073252
#>  [35,] 1.3591217 5.9234719
#>  [36,] 1.3120291 5.9705645
#>  [37,] 2.2765909 5.0060027
#>  [38,] 2.0063934 5.2762002
#>  [39,] 0.0000000 7.2825936
#>  [40,] 0.6242054 6.6583883
#>  [41,] 0.0000000 7.2825936
#>  [42,] 0.6242054 6.6583883
#>  [43,] 2.3045453 4.9780484
#>  [44,] 2.6248973 4.6576963
#>  [45,] 2.6248973 4.6576963
#>  [46,] 0.5771127 6.7054809
#>  [47,] 1.9144521 5.3681416
#>  [48,] 1.3403521 5.9422416
#>  [49,] 0.0000000 7.2825936
#>  [50,] 0.7161467 6.5664469
#>  [51,] 0.7131340 6.5694597
#>  [52,] 1.2013181 6.0812755
#>  [53,] 2.1520763 5.1305173
#>  [54,] 2.1520763 5.1305173
#>  [55,] 2.0290149 5.2535787
#>  [56,] 1.4630449 5.8195488
#>  [57,] 2.4279753 4.8546183
#>  [58,] 1.4048096 5.8777841
#>  [59,] 2.0722557 5.2103380
#>  [60,] 1.4480503 5.8345433
#>  [61,] 1.3591217 5.9234719
#>  [62,] 2.2518720 5.0307216
#>  [63,] 0.7349164 6.5476773
#>  [64,] 1.3128682 5.9697254
#>  [65,] 0.7161467 6.5664469
#>  [66,] 2.8545428 4.4280509
#>  [67,] 2.1205877 5.1620060
#>  [68,] 2.9434714 4.3391223
#>  [69,] 1.2657756 6.0168181
#>  [70,] 5.8466640 1.4359296
#>  [71,] 5.2424360 2.0401576
#>  [72,] 5.7468683 1.5357253
#>  [73,] 4.8285600 2.4540336
#>  [74,] 6.3180318 0.9645618
#>  [75,] 4.8040889 2.4785047
#>  [76,] 4.9996016 2.2829921
#>  [77,] 5.1418012 2.1407924
#>  [78,] 4.9401101 2.3424835
#>  [79,] 5.7345179 1.5480757
#>  [80,] 5.1915380 2.0910557
#>  [81,] 6.5694597 0.7131340
#>  [82,] 5.9923469 1.2902467
#>  [83,] 7.2825936 0.0000000
#>  [84,] 4.4753913 2.8072024
#>  [85,] 6.5694597 0.7131340
#>  [86,] 7.2825936 0.0000000
#>  [87,] 5.5172229 1.7653708
#>  [88,] 6.4817847 0.8008090
#>  [89,] 5.1888938 2.0936998
#>  [90,] 6.3180318 0.9645618
#>  [91,] 5.7345179 1.5480757
#>  [92,] 5.7345179 1.5480757
#>  [93,] 5.3010399 1.9815537
#>  [94,] 5.8781527 1.4044410
#>  [95,] 5.0773437 2.2052499
#>  [96,] 6.5694597 0.7131340
#>  [97,] 6.3180318 0.9645618
#>  [98,] 6.5668155 0.7157781
#>  [99,] 7.2825936 0.0000000
#> [100,] 6.4817847 0.8008090
#> [101,] 4.4604270 2.8221666
#> [102,] 7.2825936 0.0000000
#> [103,] 5.8195488 1.4630449
#> [104,] 6.5668155 0.7157781
#> [105,] 5.0310902 2.2515034
#> [106,] 4.8014448 2.4811489
#> [107,] 6.3180318 0.9645618
#> [108,] 5.9897028 1.2928909
#> [109,] 6.5939308 0.6886628
#> [110,] 5.8466640 1.4359296
#> [111,] 5.8466640 1.4359296
#> [112,] 5.8466640 1.4359296
#> [113,] 5.8536816 1.4289121
#> [114,] 5.2762002 2.0063934
#> [115,] 5.0247724 2.2578213
#> [116,] 5.1953434 2.0872503
#> [117,] 5.8466640 1.4359296
#> [118,] 5.7660066 1.5165871
#> [119,] 5.8590144 1.4235792
#> [120,] 5.2735561 2.0090376
#> [121,] 5.9893342 1.2932595
#> [122,] 4.4753913 2.8072024
#> [123,] 6.6583883 0.6242054
#> [124,] 5.9426102 1.3399835
#> [125,] 6.5668155 0.7157781
#> [126,] 6.5668155 0.7157781
#> [127,] 6.5668155 0.7157781
#> [128,] 4.6302125 2.6523812
#> [129,] 6.5694597 0.7131340
#> [130,] 6.5668155 0.7157781
#> [131,] 5.0799879 2.2026058
#> [132,] 5.2544178 2.0281758
#> [133,] 7.2825936 0.0000000
#> [134,] 6.6583883 0.6242054
#> [135,] 7.2825936 0.0000000
#> [136,] 6.6583883 0.6242054
#> [137,] 7.2825936 0.0000000
#> [138,] 5.3006714 1.9819223
#> [139,] 5.7468683 1.5357253
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.08571196 0.91428804
#>   [2,] 0.20053306 0.79946694
#>   [3,] 0.38458178 0.61541822
#>   [4,] 0.09833677 0.90166323
#>   [5,] 0.20089613 0.79910387
#>   [6,] 0.33334370 0.66665630
#>   [7,] 0.17716857 0.82283143
#>   [8,] 0.20094675 0.79905325
#>   [9,] 0.21815952 0.78184048
#>  [10,] 0.18363503 0.81636497
#>  [11,] 0.31608260 0.68391740
#>  [12,] 0.40001500 0.59998500
#>  [13,] 0.18363503 0.81636497
#>  [14,] 0.29717394 0.70282606
#>  [15,] 0.29340003 0.70659997
#>  [16,] 0.40708538 0.59291462
#>  [17,] 0.28496282 0.71503718
#>  [18,] 0.09792308 0.90207692
#>  [19,] 0.46249447 0.53750553
#>  [20,] 0.38453117 0.61546883
#>  [21,] 0.31644568 0.68355432
#>  [22,] 0.31608260 0.68391740
#>  [23,] 0.10260998 0.89739002
#>  [24,] 0.42363334 0.57636666
#>  [25,] 0.29359709 0.70640291
#>  [26,] 0.30625592 0.69374408
#>  [27,] 0.16495745 0.83504255
#>  [28,] 0.18662605 0.81337395
#>  [29,] 0.18363503 0.81636497
#>  [30,] 0.18662605 0.81337395
#>  [31,] 0.27173143 0.72826857
#>  [32,] 0.09792308 0.90207692
#>  [33,] 0.08571196 0.91428804
#>  [34,] 0.28496282 0.71503718
#>  [35,] 0.18662605 0.81337395
#>  [36,] 0.18015959 0.81984041
#>  [37,] 0.31260716 0.68739284
#>  [38,] 0.27550534 0.72449466
#>  [39,] 0.00000000 1.00000000
#>  [40,] 0.08571196 0.91428804
#>  [41,] 0.00000000 1.00000000
#>  [42,] 0.08571196 0.91428804
#>  [43,] 0.31644568 0.68355432
#>  [44,] 0.36043440 0.63956560
#>  [45,] 0.36043440 0.63956560
#>  [46,] 0.07924550 0.92075450
#>  [47,] 0.26288053 0.73711947
#>  [48,] 0.18404872 0.81595128
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.09833677 0.90166323
#>  [51,] 0.09792308 0.90207692
#>  [52,] 0.16495745 0.83504255
#>  [53,] 0.29550960 0.70449040
#>  [54,] 0.29550960 0.70449040
#>  [55,] 0.27861158 0.72138842
#>  [56,] 0.20089613 0.79910387
#>  [57,] 0.33339431 0.66660569
#>  [58,] 0.19289962 0.80710038
#>  [59,] 0.28454913 0.71545087
#>  [60,] 0.19883717 0.80116283
#>  [61,] 0.18662605 0.81337395
#>  [62,] 0.30921292 0.69078708
#>  [63,] 0.10091410 0.89908590
#>  [64,] 0.18027481 0.81972519
#>  [65,] 0.09833677 0.90166323
#>  [66,] 0.39196788 0.60803212
#>  [67,] 0.29118577 0.70881423
#>  [68,] 0.40417900 0.59582100
#>  [69,] 0.17380835 0.82619165
#>  [70,] 0.80282717 0.19717283
#>  [71,] 0.71985837 0.28014163
#>  [72,] 0.78912385 0.21087615
#>  [73,] 0.66302752 0.33697248
#>  [74,] 0.86755243 0.13244757
#>  [75,] 0.65966730 0.34033270
#>  [76,] 0.68651387 0.31348613
#>  [77,] 0.70603984 0.29396016
#>  [78,] 0.67834488 0.32165512
#>  [79,] 0.78742797 0.21257203
#>  [80,] 0.71286937 0.28713063
#>  [81,] 0.90207692 0.09792308
#>  [82,] 0.82283143 0.17716857
#>  [83,] 1.00000000 0.00000000
#>  [84,] 0.61453261 0.38546739
#>  [85,] 0.90207692 0.09792308
#>  [86,] 1.00000000 0.00000000
#>  [87,] 0.75759038 0.24240962
#>  [88,] 0.89003795 0.10996205
#>  [89,] 0.71250630 0.28749370
#>  [90,] 0.86755243 0.13244757
#>  [91,] 0.78742797 0.21257203
#>  [92,] 0.78742797 0.21257203
#>  [93,] 0.72790550 0.27209450
#>  [94,] 0.80715099 0.19284901
#>  [95,] 0.69718894 0.30281106
#>  [96,] 0.90207692 0.09792308
#>  [97,] 0.86755243 0.13244757
#>  [98,] 0.90171385 0.09828615
#>  [99,] 1.00000000 0.00000000
#> [100,] 0.89003795 0.10996205
#> [101,] 0.61247781 0.38752219
#> [102,] 1.00000000 0.00000000
#> [103,] 0.79910387 0.20089613
#> [104,] 0.90171385 0.09828615
#> [105,] 0.69083770 0.30916230
#> [106,] 0.65930423 0.34069577
#> [107,] 0.86755243 0.13244757
#> [108,] 0.82246835 0.17753165
#> [109,] 0.90543715 0.09456285
#> [110,] 0.80282717 0.19717283
#> [111,] 0.80282717 0.19717283
#> [112,] 0.80282717 0.19717283
#> [113,] 0.80379077 0.19620923
#> [114,] 0.72449466 0.27550534
#> [115,] 0.68997017 0.31002983
#> [116,] 0.71339191 0.28660809
#> [117,] 0.80282717 0.19717283
#> [118,] 0.79175179 0.20824821
#> [119,] 0.80452305 0.19547695
#> [120,] 0.72413159 0.27586841
#> [121,] 0.82241774 0.17758226
#> [122,] 0.61453261 0.38546739
#> [123,] 0.91428804 0.08571196
#> [124,] 0.81600189 0.18399811
#> [125,] 0.90171385 0.09828615
#> [126,] 0.90171385 0.09828615
#> [127,] 0.90171385 0.09828615
#> [128,] 0.63579168 0.36420832
#> [129,] 0.90207692 0.09792308
#> [130,] 0.90171385 0.09828615
#> [131,] 0.69755201 0.30244799
#> [132,] 0.72150364 0.27849636
#> [133,] 1.00000000 0.00000000
#> [134,] 0.91428804 0.08571196
#> [135,] 1.00000000 0.00000000
#> [136,] 0.91428804 0.08571196
#> [137,] 1.00000000 0.00000000
#> [138,] 0.72785488 0.27214512
#> [139,] 0.78912385 0.21087615
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>        V1       V10       V11       V12       V13       V14       V15       V16 
#> 0.0000000 2.3719341 8.5295442 7.6469582 0.0000000 0.0000000 1.0096362 0.6258345 
#>       V17       V18       V19        V2       V20       V21       V22       V23 
#> 0.0000000 3.7136494 1.7657674 2.7079642 1.1766865 2.5007232 0.9002101 3.8378254 
#>       V24       V25       V26       V27       V28       V29        V3       V30 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 2.2890786 0.0000000 
#>       V31       V32       V33       V34       V35       V36       V37       V38 
#> 2.1762753 0.0000000 1.8370944 0.0000000 5.8244002 2.3309750 1.1001794 0.0000000 
#>       V39        V4       V40       V41       V42       V43       V44       V45 
#> 0.0000000 4.0061817 0.0000000 0.0000000 0.0000000 2.1044761 0.0000000 2.2104797 
#>       V46       V47       V48       V49        V5       V50       V51       V52 
#> 1.3168995 0.0000000 0.0000000 6.9252140 0.9743271 0.0000000 0.0000000 9.9376409 
#>       V53       V54       V55       V56       V57       V58       V59        V6 
#> 1.6335787 1.4338866 3.3019753 0.8559896 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V60        V7        V8        V9 
#> 0.0000000 4.6965325 1.1344991 7.1235831 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 70 69 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>       V52       V11       V12        V9       V49       V35        V7        V4 
#> 9.9376409 8.5295442 7.6469582 7.1235831 6.9252140 5.8244002 4.6965325 4.0061817 
#>       V23       V18       V55        V2       V21       V10       V36        V3 
#> 3.8378254 3.7136494 3.3019753 2.7079642 2.5007232 2.3719341 2.3309750 2.2890786 
#>       V45       V31       V43       V33       V19       V53       V54       V46 
#> 2.2104797 2.1762753 2.1044761 1.8370944 1.7657674 1.6335787 1.4338866 1.3168995 
#>       V20        V8       V37       V15        V5       V22       V56       V16 
#> 1.1766865 1.1344991 1.1001794 1.0096362 0.9743271 0.9002101 0.8559896 0.6258345 
#>        V1       V13       V14       V17       V24       V25       V26       V27 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V28       V29       V30       V32       V34       V38       V39       V40 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V41       V42       V44       V47       V48       V50       V51       V57 
#> 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 
#>       V58       V59        V6       V60 
#> 0.0000000 0.0000000 0.0000000 0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058