Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V49>=0.03845 77 19 M (0.75324675 0.24675325)  
#>      4) V31< 0.4599 31  0 M (1.00000000 0.00000000) *
#>      5) V31>=0.4599 46 19 M (0.58695652 0.41304348)  
#>       10) V28>=0.66175 29  2 M (0.93103448 0.06896552) *
#>       11) V28< 0.66175 17  0 R (0.00000000 1.00000000) *
#>    3) V49< 0.03845 62 17 R (0.27419355 0.72580645)  
#>      6) V11>=0.28665 17  4 M (0.76470588 0.23529412) *
#>      7) V11< 0.28665 45  4 R (0.08888889 0.91111111) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.46043165 0.53956835)  
#>    2) V11>=0.13425 109 46 M (0.57798165 0.42201835)  
#>      4) V16< 0.39975 48  8 M (0.83333333 0.16666667)  
#>        8) V10>=0.15185 40  1 M (0.97500000 0.02500000) *
#>        9) V10< 0.15185 8  1 R (0.12500000 0.87500000) *
#>      5) V16>=0.39975 61 23 R (0.37704918 0.62295082)  
#>       10) V21>=0.8191 14  0 M (1.00000000 0.00000000) *
#>       11) V21< 0.8191 47  9 R (0.19148936 0.80851064) *
#>    3) V11< 0.13425 30  1 R (0.03333333 0.96666667) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 57 M (0.58992806 0.41007194)  
#>    2) V5>=0.0392 107 29 M (0.72897196 0.27102804)  
#>      4) V4>=0.03795 76 10 M (0.86842105 0.13157895)  
#>        8) V30>=0.46225 51  1 M (0.98039216 0.01960784) *
#>        9) V30< 0.46225 25  9 M (0.64000000 0.36000000)  
#>         18) V27< 0.56135 16  0 M (1.00000000 0.00000000) *
#>         19) V27>=0.56135 9  0 R (0.00000000 1.00000000) *
#>      5) V4< 0.03795 31 12 R (0.38709677 0.61290323)  
#>       10) V35< 0.22685 12  2 M (0.83333333 0.16666667) *
#>       11) V35>=0.22685 19  2 R (0.10526316 0.89473684) *
#>    3) V5< 0.0392 32  4 R (0.12500000 0.87500000)  
#>      6) V1>=0.0228 7  3 M (0.57142857 0.42857143) *
#>      7) V1< 0.0228 25  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V4>=0.052 52  8 M (0.84615385 0.15384615)  
#>      4) V5>=0.03665 45  2 M (0.95555556 0.04444444) *
#>      5) V5< 0.03665 7  1 R (0.14285714 0.85714286) *
#>    3) V4< 0.052 87 34 R (0.39080460 0.60919540)  
#>      6) V21>=0.67375 38 12 M (0.68421053 0.31578947)  
#>       12) V20< 0.84235 24  0 M (1.00000000 0.00000000) *
#>       13) V20>=0.84235 14  2 R (0.14285714 0.85714286) *
#>      7) V21< 0.67375 49  8 R (0.16326531 0.83673469)  
#>       14) V36< 0.15855 12  4 M (0.66666667 0.33333333) *
#>       15) V36>=0.15855 37  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V46>=0.06695 103 40 M (0.61165049 0.38834951)  
#>      4) V12>=0.1677 65 11 M (0.83076923 0.16923077)  
#>        8) V27>=0.71565 44  1 M (0.97727273 0.02272727) *
#>        9) V27< 0.71565 21 10 M (0.52380952 0.47619048)  
#>         18) V31< 0.4195 10  0 M (1.00000000 0.00000000) *
#>         19) V31>=0.4195 11  1 R (0.09090909 0.90909091) *
#>      5) V12< 0.1677 38  9 R (0.23684211 0.76315789)  
#>       10) V25< 0.45365 11  2 M (0.81818182 0.18181818) *
#>       11) V25>=0.45365 27  0 R (0.00000000 1.00000000) *
#>    3) V46< 0.06695 36  3 R (0.08333333 0.91666667) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 R (0.47482014 0.52517986)  
#>    2) V4>=0.0582 54 13 M (0.75925926 0.24074074)  
#>      4) V50>=0.00645 46  6 M (0.86956522 0.13043478)  
#>        8) V34>=0.1883 39  0 M (1.00000000 0.00000000) *
#>        9) V34< 0.1883 7  1 R (0.14285714 0.85714286) *
#>      5) V50< 0.00645 8  1 R (0.12500000 0.87500000) *
#>    3) V4< 0.0582 85 25 R (0.29411765 0.70588235)  
#>      6) V28>=0.92035 19  5 M (0.73684211 0.26315789) *
#>      7) V28< 0.92035 66 11 R (0.16666667 0.83333333)  
#>       14) V28< 0.37175 7  0 M (1.00000000 0.00000000) *
#>       15) V28>=0.37175 59  4 R (0.06779661 0.93220339) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 61 M (0.56115108 0.43884892)  
#>    2) V21>=0.6401 79 17 M (0.78481013 0.21518987)  
#>      4) V31< 0.56955 69  7 M (0.89855072 0.10144928)  
#>        8) V44>=0.1384 56  0 M (1.00000000 0.00000000) *
#>        9) V44< 0.1384 13  6 R (0.46153846 0.53846154) *
#>      5) V31>=0.56955 10  0 R (0.00000000 1.00000000) *
#>    3) V21< 0.6401 60 16 R (0.26666667 0.73333333)  
#>      6) V47>=0.31685 7  0 M (1.00000000 0.00000000) *
#>      7) V47< 0.31685 53  9 R (0.16981132 0.83018868)  
#>       14) V36< 0.2083 8  2 M (0.75000000 0.25000000) *
#>       15) V36>=0.2083 45  3 R (0.06666667 0.93333333) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 62 M (0.55395683 0.44604317)  
#>    2) V51>=0.01285 79 21 M (0.73417722 0.26582278)  
#>      4) V36< 0.44265 53  4 M (0.92452830 0.07547170) *
#>      5) V36>=0.44265 26  9 R (0.34615385 0.65384615)  
#>       10) V12>=0.24295 11  3 M (0.72727273 0.27272727) *
#>       11) V12< 0.24295 15  1 R (0.06666667 0.93333333) *
#>    3) V51< 0.01285 60 19 R (0.31666667 0.68333333)  
#>      6) V44>=0.2503 12  0 M (1.00000000 0.00000000) *
#>      7) V44< 0.2503 48  7 R (0.14583333 0.85416667)  
#>       14) V35< 0.2434 20  7 R (0.35000000 0.65000000)  
#>         28) V52>=0.00795 8  2 M (0.75000000 0.25000000) *
#>         29) V52< 0.00795 12  1 R (0.08333333 0.91666667) *
#>       15) V35>=0.2434 28  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 55 M (0.60431655 0.39568345)  
#>   2) V21>=0.68525 76  9 M (0.88157895 0.11842105)  
#>     4) V18< 0.9179 68  4 M (0.94117647 0.05882353) *
#>     5) V18>=0.9179 8  3 R (0.37500000 0.62500000) *
#>   3) V21< 0.68525 63 17 R (0.26984127 0.73015873)  
#>     6) V48>=0.0824 19  2 M (0.89473684 0.10526316) *
#>     7) V48< 0.0824 44  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 60 M (0.56834532 0.43165468)  
#>    2) V11>=0.1168 107 29 M (0.72897196 0.27102804)  
#>      4) V52>=0.0088 69  8 M (0.88405797 0.11594203)  
#>        8) V12>=0.14275 62  4 M (0.93548387 0.06451613) *
#>        9) V12< 0.14275 7  3 R (0.42857143 0.57142857) *
#>      5) V52< 0.0088 38 17 R (0.44736842 0.55263158)  
#>       10) V23>=0.82995 12  0 M (1.00000000 0.00000000) *
#>       11) V23< 0.82995 26  5 R (0.19230769 0.80769231)  
#>         22) V43>=0.24095 7  2 M (0.71428571 0.28571429) *
#>         23) V43< 0.24095 19  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.1168 32  1 R (0.03125000 0.96875000) *
#> 
#> 
#> $weights
#>  [1] 0.7113310 0.8188278 0.6528871 0.7257103 0.9020586 0.6764200 0.9782929
#>  [8] 0.6074625 0.7384970 0.9923752
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.5497000 5.2541624
#>   [2,] 2.6973910 5.1064714
#>   [3,] 2.0223795 5.7814829
#>   [4,] 2.6481006 5.1557618
#>   [5,] 1.3642181 6.4396442
#>   [6,] 1.4642073 6.3396551
#>   [7,] 2.6188485 5.1850139
#>   [8,] 2.3518866 5.4519758
#>   [9,] 2.3114657 5.4923967
#>  [10,] 2.0716698 5.7321925
#>  [11,] 2.0223795 5.7814829
#>  [12,] 1.5095211 6.2943413
#>  [13,] 1.9860600 5.8178024
#>  [14,] 2.3383347 5.4655277
#>  [15,] 1.4149169 6.3889454
#>  [16,] 0.6074625 7.1963999
#>  [17,] 2.3568903 5.4469720
#>  [18,] 1.7167899 6.0870725
#>  [19,] 2.5473209 5.2565415
#>  [20,] 0.0000000 7.8038624
#>  [21,] 0.7384970 7.0653654
#>  [22,] 1.3913841 6.4124783
#>  [23,] 2.1027151 5.7011472
#>  [24,] 3.1443758 4.6594866
#>  [25,] 0.8188278 6.9850346
#>  [26,] 0.0000000 7.8038624
#>  [27,] 0.9020586 6.9018038
#>  [28,] 3.1783163 4.6255461
#>  [29,] 1.4149169 6.3889454
#>  [30,] 1.9706681 5.8331943
#>  [31,] 0.0000000 7.8038624
#>  [32,] 0.6528871 7.1509752
#>  [33,] 2.2527248 5.5511375
#>  [34,] 2.0406381 5.7632243
#>  [35,] 3.0283065 4.7755558
#>  [36,] 0.9923752 6.8114872
#>  [37,] 0.0000000 7.8038624
#>  [38,] 1.6452623 6.1586001
#>  [39,] 2.5369133 5.2669491
#>  [40,] 1.3293071 6.4745553
#>  [41,] 2.1027151 5.7011472
#>  [42,] 0.7113310 7.0925314
#>  [43,] 1.5998377 6.2040247
#>  [44,] 0.0000000 7.8038624
#>  [45,] 0.9782929 6.8255695
#>  [46,] 2.2386425 5.5652199
#>  [47,] 1.3459595 6.4579029
#>  [48,] 2.7101776 5.0936847
#>  [49,] 0.8188278 6.9850346
#>  [50,] 2.9784352 4.8254272
#>  [51,] 1.3785975 6.4252649
#>  [52,] 1.3293071 6.4745553
#>  [53,] 0.8188278 6.9850346
#>  [54,] 2.1755383 5.6283241
#>  [55,] 2.4422032 5.3616592
#>  [56,] 2.1027151 5.7011472
#>  [57,] 2.1755383 5.6283241
#>  [58,] 1.3913841 6.4124783
#>  [59,] 1.3187935 6.4850689
#>  [60,] 0.9923752 6.8114872
#>  [61,] 2.9643529 4.8395095
#>  [62,] 1.7040032 6.0998592
#>  [63,] 1.3459595 6.4579029
#>  [64,] 5.5416870 2.2621753
#>  [65,] 6.9850346 0.8188278
#>  [66,] 4.6680590 3.1358033
#>  [67,] 5.1046831 2.6991793
#>  [68,] 6.2943413 1.5095211
#>  [69,] 5.3851969 2.4186655
#>  [70,] 6.9850346 0.8188278
#>  [71,] 5.6904011 2.1134613
#>  [72,] 6.5435127 1.2603496
#>  [73,] 7.0925314 0.7113310
#>  [74,] 6.4017321 1.4021303
#>  [75,] 6.1904728 1.6133896
#>  [76,] 6.4017321 1.4021303
#>  [77,] 7.1274424 0.6764200
#>  [78,] 7.8038624 0.0000000
#>  [79,] 6.9850346 0.8188278
#>  [80,] 7.1274424 0.6764200
#>  [81,] 7.1274424 0.6764200
#>  [82,] 5.3303217 2.4735406
#>  [83,] 6.4017321 1.4021303
#>  [84,] 5.5724967 2.2313657
#>  [85,] 7.0781520 0.7257103
#>  [86,] 6.9850346 0.8188278
#>  [87,] 6.2040247 1.5998377
#>  [88,] 6.2040247 1.5998377
#>  [89,] 6.3775721 1.4262903
#>  [90,] 6.2040247 1.5998377
#>  [91,] 6.2943413 1.5095211
#>  [92,] 6.1760935 1.6277689
#>  [93,] 5.3716450 2.4322174
#>  [94,] 6.1760935 1.6277689
#>  [95,] 5.9926594 1.8112030
#>  [96,] 6.0067417 1.7971207
#>  [97,] 7.1274424 0.6764200
#>  [98,] 5.2121799 2.5916825
#>  [99,] 5.1978006 2.6060618
#> [100,] 4.7641825 3.0396799
#> [101,] 4.5556181 3.2482443
#> [102,] 6.8255695 0.9782929
#> [103,] 6.8255695 0.9782929
#> [104,] 5.4234392 2.3804232
#> [105,] 5.7488450 2.0550174
#> [106,] 6.1350672 1.6687951
#> [107,] 5.6208164 2.1830459
#> [108,] 5.5352066 2.2686558
#> [109,] 6.2181070 1.5857554
#> [110,] 7.1274424 0.6764200
#> [111,] 7.1274424 0.6764200
#> [112,] 5.8670928 1.9367696
#> [113,] 5.7011472 2.1027151
#> [114,] 7.8038624 0.0000000
#> [115,] 7.1274424 0.6764200
#> [116,] 7.8038624 0.0000000
#> [117,] 6.8255695 0.9782929
#> [118,] 7.8038624 0.0000000
#> [119,] 7.8038624 0.0000000
#> [120,] 6.2253838 1.5784786
#> [121,] 7.8038624 0.0000000
#> [122,] 7.1509752 0.6528871
#> [123,] 7.8038624 0.0000000
#> [124,] 7.0925314 0.7113310
#> [125,] 6.8114872 0.9923752
#> [126,] 7.0653654 0.7384970
#> [127,] 7.8038624 0.0000000
#> [128,] 7.8038624 0.0000000
#> [129,] 7.8038624 0.0000000
#> [130,] 7.8038624 0.0000000
#> [131,] 7.1509752 0.6528871
#> [132,] 7.8038624 0.0000000
#> [133,] 7.8038624 0.0000000
#> [134,] 6.9018038 0.9020586
#> [135,] 7.8038624 0.0000000
#> [136,] 6.2943413 1.5095211
#> [137,] 7.8038624 0.0000000
#> [138,] 5.4093569 2.3945055
#> [139,] 4.9519041 2.8519583
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32672283 0.67327717
#>   [2,] 0.34564820 0.65435180
#>   [3,] 0.25915109 0.74084891
#>   [4,] 0.33933205 0.66066795
#>   [5,] 0.17481320 0.82518680
#>   [6,] 0.18762598 0.81237402
#>   [7,] 0.33558363 0.66441637
#>   [8,] 0.30137469 0.69862531
#>   [9,] 0.29619509 0.70380491
#>  [10,] 0.26546724 0.73453276
#>  [11,] 0.25915109 0.74084891
#>  [12,] 0.19343256 0.80656744
#>  [13,] 0.25449705 0.74550295
#>  [14,] 0.29963812 0.70036188
#>  [15,] 0.18130983 0.81869017
#>  [16,] 0.07784126 0.92215874
#>  [17,] 0.30201588 0.69798412
#>  [18,] 0.21999233 0.78000767
#>  [19,] 0.32641797 0.67358203
#>  [20,] 0.00000000 1.00000000
#>  [21,] 0.09463224 0.90536776
#>  [22,] 0.17829429 0.82170571
#>  [23,] 0.26944544 0.73055456
#>  [24,] 0.40292559 0.59707441
#>  [25,] 0.10492597 0.89507403
#>  [26,] 0.00000000 1.00000000
#>  [27,] 0.11559130 0.88440870
#>  [28,] 0.40727477 0.59272523
#>  [29,] 0.18130983 0.81869017
#>  [30,] 0.25252471 0.74747529
#>  [31,] 0.00000000 1.00000000
#>  [32,] 0.08366205 0.91633795
#>  [33,] 0.28866793 0.71133207
#>  [34,] 0.26149078 0.73850922
#>  [35,] 0.38805227 0.61194773
#>  [36,] 0.12716462 0.87283538
#>  [37,] 0.00000000 1.00000000
#>  [38,] 0.21082667 0.78917333
#>  [39,] 0.32508432 0.67491568
#>  [40,] 0.17033964 0.82966036
#>  [41,] 0.26944544 0.73055456
#>  [42,] 0.09115115 0.90884885
#>  [43,] 0.20500588 0.79499412
#>  [44,] 0.00000000 1.00000000
#>  [45,] 0.12536009 0.87463991
#>  [46,] 0.28686340 0.71313660
#>  [47,] 0.17247351 0.82752649
#>  [48,] 0.34728670 0.65271330
#>  [49,] 0.10492597 0.89507403
#>  [50,] 0.38166167 0.61833833
#>  [51,] 0.17665579 0.82334421
#>  [52,] 0.17033964 0.82966036
#>  [53,] 0.10492597 0.89507403
#>  [54,] 0.27877713 0.72122287
#>  [55,] 0.31294801 0.68705199
#>  [56,] 0.26944544 0.73055456
#>  [57,] 0.27877713 0.72122287
#>  [58,] 0.17829429 0.82170571
#>  [59,] 0.16899241 0.83100759
#>  [60,] 0.12716462 0.87283538
#>  [61,] 0.37985714 0.62014286
#>  [62,] 0.21835383 0.78164617
#>  [63,] 0.17247351 0.82752649
#>  [64,] 0.71012106 0.28987894
#>  [65,] 0.89507403 0.10492597
#>  [66,] 0.59817291 0.40182709
#>  [67,] 0.65412265 0.34587735
#>  [68,] 0.80656744 0.19343256
#>  [69,] 0.69006815 0.30993185
#>  [70,] 0.89507403 0.10492597
#>  [71,] 0.72917753 0.27082247
#>  [72,] 0.83849668 0.16150332
#>  [73,] 0.90884885 0.09115115
#>  [74,] 0.82032868 0.17967132
#>  [75,] 0.79325755 0.20674245
#>  [76,] 0.82032868 0.17967132
#>  [77,] 0.91332241 0.08667759
#>  [78,] 1.00000000 0.00000000
#>  [79,] 0.89507403 0.10492597
#>  [80,] 0.91332241 0.08667759
#>  [81,] 0.91332241 0.08667759
#>  [82,] 0.68303636 0.31696364
#>  [83,] 0.82032868 0.17967132
#>  [84,] 0.71406906 0.28593094
#>  [85,] 0.90700626 0.09299374
#>  [86,] 0.89507403 0.10492597
#>  [87,] 0.79499412 0.20500588
#>  [88,] 0.79499412 0.20500588
#>  [89,] 0.81723277 0.18276723
#>  [90,] 0.79499412 0.20500588
#>  [91,] 0.80656744 0.19343256
#>  [92,] 0.79141496 0.20858504
#>  [93,] 0.68833159 0.31166841
#>  [94,] 0.79141496 0.20858504
#>  [95,] 0.76790942 0.23209058
#>  [96,] 0.76971395 0.23028605
#>  [97,] 0.91332241 0.08667759
#>  [98,] 0.66789747 0.33210253
#>  [99,] 0.66605487 0.33394513
#> [100,] 0.61049032 0.38950968
#> [101,] 0.58376453 0.41623547
#> [102,] 0.87463991 0.12536009
#> [103,] 0.87463991 0.12536009
#> [104,] 0.69496859 0.30503141
#> [105,] 0.73666662 0.26333338
#> [106,] 0.78615779 0.21384221
#> [107,] 0.72026084 0.27973916
#> [108,] 0.70929065 0.29070935
#> [109,] 0.79679865 0.20320135
#> [110,] 0.91332241 0.08667759
#> [111,] 0.91332241 0.08667759
#> [112,] 0.75181910 0.24818090
#> [113,] 0.73055456 0.26944544
#> [114,] 1.00000000 0.00000000
#> [115,] 0.91332241 0.08667759
#> [116,] 1.00000000 0.00000000
#> [117,] 0.87463991 0.12536009
#> [118,] 1.00000000 0.00000000
#> [119,] 1.00000000 0.00000000
#> [120,] 0.79773111 0.20226889
#> [121,] 1.00000000 0.00000000
#> [122,] 0.91633795 0.08366205
#> [123,] 1.00000000 0.00000000
#> [124,] 0.90884885 0.09115115
#> [125,] 0.87283538 0.12716462
#> [126,] 0.90536776 0.09463224
#> [127,] 1.00000000 0.00000000
#> [128,] 1.00000000 0.00000000
#> [129,] 1.00000000 0.00000000
#> [130,] 1.00000000 0.00000000
#> [131,] 0.91633795 0.08366205
#> [132,] 1.00000000 0.00000000
#> [133,] 1.00000000 0.00000000
#> [134,] 0.88440870 0.11559130
#> [135,] 1.00000000 0.00000000
#> [136,] 0.80656744 0.19343256
#> [137,] 1.00000000 0.00000000
#> [138,] 0.69316406 0.30683594
#> [139,] 0.63454529 0.36545471
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  0.5675397  1.9199299 10.5260890  5.3142941  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  2.2292261  0.0000000  0.8248669  0.0000000  0.0000000  2.2949414 13.8565511 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  0.0000000  2.5875275  0.0000000  2.2974890  0.0000000  3.1142383  6.5867490 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.6177887  6.3533168  0.0000000  0.0000000  1.4357335 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  1.6617117  4.6586182  0.0000000  0.0000000  0.0000000  6.3577634  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  0.0000000  0.0000000  1.2607945  3.5280234  0.0000000  3.2696050  2.0295108 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  3.8190144  2.7293111  4.2692934  1.2440051  1.7577784  2.8882895  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 76 63 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V21        V11        V28         V4        V31        V12        V36 
#> 13.8565511 10.5260890  6.5867490  6.3577634  6.3533168  5.3142941  4.6586182 
#>         V5        V48        V44        V46        V27        V52        V49 
#>  4.2692934  3.8190144  3.5280234  3.2696050  3.1142383  2.8882895  2.7293111 
#>        V23        V25        V20        V16        V47        V10        V51 
#>  2.5875275  2.2974890  2.2949414  2.2292261  2.0295108  1.9199299  1.7577784 
#>        V35        V34        V43        V50        V18        V30         V1 
#>  1.6617117  1.4357335  1.2607945  1.2440051  0.8248669  0.6177887  0.5675397 
#>        V13        V14        V15        V17        V19         V2        V22 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V24        V26        V29         V3        V32        V33        V37 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V38        V39        V40        V41        V42        V45        V53 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058