Classification Boosting Learner

Classification boosting algorithm. Calls adabag::boosting() from adabag.

Initial parameter values

xval:

Actual default: 10L

Initial value: 0L

Reason for change: Set to 0 for speed.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.adabag")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”

Required Packages: mlr3, adabag, rpart

Parameters

Id	Type	Default	Levels	Range

boos	logical	TRUE	TRUE, FALSE	-

coeflearn	character	Breiman	Breiman, Freund, Zhu	-

cp	numeric	0.01		\([0, 1]\)

maxcompete	integer	4		\([0, \infty)\)

maxdepth	integer	30		\([1, 30]\)

maxsurrogate	integer	5		\([0, \infty)\)

mfinal	integer	100		\([1, \infty)\)

minbucket	integer	-		\([1, \infty)\)

minsplit	integer	20		\([1, \infty)\)

newmfinal	integer	-		\((-\infty, \infty)\)

surrogatestyle	integer	0		\([0, 1]\)

usesurrogate	integer	2		\([0, 2]\)

xval	integer	0		\([0, \infty)\)

References

Alfaro, Esteban, Gamez, Matias, García, Noelia (2013). “adabag: An R Package for Classification with Boosting and Bagging.” Journal of Statistical Software, 54(2), 1-35. doi:10.18637/jss.v054.i02 , https://www.jstatsoft.org/index.php/jss/article/view/v054i02.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifAdabag

Methods

Public methods

LearnerClassifAdabag$new()

LearnerClassifAdabag$importance()

LearnerClassifAdabag$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifAdabag$new()

Method importance()

The importance scores are extracted from the model.

Usage

LearnerClassifAdabag$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifAdabag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.adabag", mfinal = 10L)
print(learner)
#> 
#> ── <LearnerClassifAdabag> (classif.adabag): Adabag Boosting ────────────────────
#> • Model: -
#> • Parameters: mfinal=10, xval=0
#> • Packages: mlr3, adabag, and rpart
#> • Predict Types: [response] and prob
#> • Feature Types: integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $formula
#> Class ~ .
#> NULL
#> 
#> $trees
#> $trees[[1]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V11>=0.19825 81 16 M (0.80246914 0.19753086)  
#>      4) V18< 0.5913 51  3 M (0.94117647 0.05882353) *
#>      5) V18>=0.5913 30 13 M (0.56666667 0.43333333)  
#>       10) V37< 0.49325 21  4 M (0.80952381 0.19047619) *
#>       11) V37>=0.49325 9  0 R (0.00000000 1.00000000) *
#>    3) V11< 0.19825 58 11 R (0.18965517 0.81034483)  
#>      6) V54>=0.02105 8  0 M (1.00000000 0.00000000) *
#>      7) V54< 0.02105 50  3 R (0.06000000 0.94000000) *
#> 
#> $trees[[2]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 65 M (0.53237410 0.46762590)  
#>   2) V9>=0.1159 97 29 M (0.70103093 0.29896907)  
#>     4) V16< 0.67795 76 11 M (0.85526316 0.14473684)  
#>       8) V56< 0.01355 69  5 M (0.92753623 0.07246377) *
#>       9) V56>=0.01355 7  1 R (0.14285714 0.85714286) *
#>     5) V16>=0.67795 21  3 R (0.14285714 0.85714286) *
#>   3) V9< 0.1159 42  6 R (0.14285714 0.85714286)  
#>     6) V21>=0.86385 7  1 M (0.85714286 0.14285714) *
#>     7) V21< 0.86385 35  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[3]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 R (0.45323741 0.54676259)  
#>    2) V51>=0.01365 80 28 M (0.65000000 0.35000000)  
#>      4) V12>=0.1786 47  4 M (0.91489362 0.08510638) *
#>      5) V12< 0.1786 33  9 R (0.27272727 0.72727273)  
#>       10) V45>=0.24845 7  0 M (1.00000000 0.00000000) *
#>       11) V45< 0.24845 26  2 R (0.07692308 0.92307692) *
#>    3) V51< 0.01365 59 11 R (0.18644068 0.81355932) *
#> 
#> $trees[[4]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 63 M (0.54676259 0.45323741)  
#>    2) V10>=0.206 69 14 M (0.79710145 0.20289855)  
#>      4) V48>=0.04995 56  5 M (0.91071429 0.08928571)  
#>        8) V37< 0.5297 48  0 M (1.00000000 0.00000000) *
#>        9) V37>=0.5297 8  3 R (0.37500000 0.62500000) *
#>      5) V48< 0.04995 13  4 R (0.30769231 0.69230769) *
#>    3) V10< 0.206 70 21 R (0.30000000 0.70000000)  
#>      6) V45>=0.1728 18  2 M (0.88888889 0.11111111) *
#>      7) V45< 0.1728 52  5 R (0.09615385 0.90384615)  
#>       14) V20>=0.87785 9  4 M (0.55555556 0.44444444) *
#>       15) V20< 0.87785 43  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[5]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 139 55 M (0.60431655 0.39568345)  
#>   2) V9>=0.1164 100 18 M (0.82000000 0.18000000)  
#>     4) V8>=0.0906 86  8 M (0.90697674 0.09302326)  
#>       8) V46>=0.0599 79  3 M (0.96202532 0.03797468) *
#>       9) V46< 0.0599 7  2 R (0.28571429 0.71428571) *
#>     5) V8< 0.0906 14  4 R (0.28571429 0.71428571) *
#>   3) V9< 0.1164 39  2 R (0.05128205 0.94871795) *
#> 
#> $trees[[6]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 M (0.53956835 0.46043165)  
#>    2) V11>=0.3049 35  0 M (1.00000000 0.00000000) *
#>    3) V11< 0.3049 104 40 R (0.38461538 0.61538462)  
#>      6) V28>=0.9576 11  0 M (1.00000000 0.00000000) *
#>      7) V28< 0.9576 93 29 R (0.31182796 0.68817204)  
#>       14) V22>=0.94265 11  1 M (0.90909091 0.09090909) *
#>       15) V22< 0.94265 82 19 R (0.23170732 0.76829268)  
#>         30) V55>=0.0206 11  1 M (0.90909091 0.09090909) *
#>         31) V55< 0.0206 71  9 R (0.12676056 0.87323944)  
#>           62) V60< 0.00215 8  3 M (0.62500000 0.37500000) *
#>           63) V60>=0.00215 63  4 R (0.06349206 0.93650794) *
#> 
#> $trees[[7]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 59 M (0.57553957 0.42446043)  
#>    2) V11>=0.2044 94 24 M (0.74468085 0.25531915)  
#>      4) V16< 0.67795 63  5 M (0.92063492 0.07936508)  
#>        8) V20< 0.92945 56  1 M (0.98214286 0.01785714) *
#>        9) V20>=0.92945 7  3 R (0.42857143 0.57142857) *
#>      5) V16>=0.67795 31 12 R (0.38709677 0.61290323)  
#>       10) V41< 0.1606 11  0 M (1.00000000 0.00000000) *
#>       11) V41>=0.1606 20  1 R (0.05000000 0.95000000) *
#>    3) V11< 0.2044 45 10 R (0.22222222 0.77777778)  
#>      6) V45>=0.16285 20 10 M (0.50000000 0.50000000)  
#>       12) V21>=0.5189 11  1 M (0.90909091 0.09090909) *
#>       13) V21< 0.5189 9  0 R (0.00000000 1.00000000) *
#>      7) V45< 0.16285 25  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[8]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 68 R (0.48920863 0.51079137)  
#>    2) V36< 0.2162 39  6 M (0.84615385 0.15384615)  
#>      4) V48>=0.03925 31  1 M (0.96774194 0.03225806) *
#>      5) V48< 0.03925 8  3 R (0.37500000 0.62500000) *
#>    3) V36>=0.2162 100 35 R (0.35000000 0.65000000)  
#>      6) V46>=0.22195 19  0 M (1.00000000 0.00000000) *
#>      7) V46< 0.22195 81 16 R (0.19753086 0.80246914)  
#>       14) V21>=0.6216 34 16 R (0.47058824 0.52941176)  
#>         28) V45< 0.12855 19  3 M (0.84210526 0.15789474) *
#>         29) V45>=0.12855 15  0 R (0.00000000 1.00000000) *
#>       15) V21< 0.6216 47  0 R (0.00000000 1.00000000) *
#> 
#> $trees[[9]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 64 R (0.4604317 0.5395683)  
#>    2) V27>=0.9009 30  4 M (0.8666667 0.1333333)  
#>      4) V58< 0.00765 23  0 M (1.0000000 0.0000000) *
#>      5) V58>=0.00765 7  3 R (0.4285714 0.5714286) *
#>    3) V27< 0.9009 109 38 R (0.3486239 0.6513761)  
#>      6) V55< 0.00335 15  2 M (0.8666667 0.1333333) *
#>      7) V55>=0.00335 94 25 R (0.2659574 0.7340426)  
#>       14) V1>=0.05225 10  1 M (0.9000000 0.1000000) *
#>       15) V1< 0.05225 84 16 R (0.1904762 0.8095238)  
#>         30) V54>=0.0183 11  3 M (0.7272727 0.2727273) *
#>         31) V54< 0.0183 73  8 R (0.1095890 0.8904110) *
#> 
#> $trees[[10]]
#> n= 139 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 139 66 M (0.52517986 0.47482014)  
#>    2) V23>=0.7845 66 14 M (0.78787879 0.21212121)  
#>      4) V23< 0.946 54  5 M (0.90740741 0.09259259)  
#>        8) V37< 0.47175 47  0 M (1.00000000 0.00000000) *
#>        9) V37>=0.47175 7  2 R (0.28571429 0.71428571) *
#>      5) V23>=0.946 12  3 R (0.25000000 0.75000000) *
#>    3) V23< 0.7845 73 21 R (0.28767123 0.71232877)  
#>      6) V4>=0.0477 21  6 M (0.71428571 0.28571429)  
#>       12) V50>=0.01225 14  0 M (1.00000000 0.00000000) *
#>       13) V50< 0.01225 7  1 R (0.14285714 0.85714286) *
#>      7) V4< 0.0477 52  6 R (0.11538462 0.88461538) *
#> 
#> 
#> $weights
#>  [1] 0.9215264 0.6302724 0.7109140 0.6284585 0.6665689 0.7486637 1.1659618
#>  [8] 0.7815315 0.5409693 1.0176108
#> 
#> $votes
#>             [,1]      [,2]
#>   [1,] 2.5434447 5.2690326
#>   [2,] 2.2986486 5.5138288
#>   [3,] 0.7486637 7.0638136
#>   [4,] 2.1209040 5.6915733
#>   [5,] 2.1290646 5.6834127
#>   [6,] 0.7486637 7.0638136
#>   [7,] 1.5517988 6.2606785
#>   [8,] 2.0402624 5.7722149
#>   [9,] 1.7991423 6.0133351
#>  [10,] 2.1604677 5.6520096
#>  [11,] 1.8803418 5.9321355
#>  [12,] 0.0000000 7.8124773
#>  [13,] 0.0000000 7.8124773
#>  [14,] 0.0000000 7.8124773
#>  [15,] 2.3811041 5.4313732
#>  [16,] 2.4628031 5.3496742
#>  [17,] 3.0037725 4.8087049
#>  [18,] 0.6284585 7.1840189
#>  [19,] 0.7486637 7.0638136
#>  [20,] 1.1694278 6.6430495
#>  [21,] 1.7991423 6.0133351
#>  [22,] 1.3789362 6.4335412
#>  [23,] 1.0176108 6.7948666
#>  [24,] 0.0000000 7.8124773
#>  [25,] 2.0909542 5.7215231
#>  [26,] 1.4924455 6.3200318
#>  [27,] 2.0783728 5.7341045
#>  [28,] 2.7177606 5.0947167
#>  [29,] 1.0176108 6.7948666
#>  [30,] 0.0000000 7.8124773
#>  [31,] 1.8768759 5.9356015
#>  [32,] 2.1209040 5.6915733
#>  [33,] 1.0176108 6.7948666
#>  [34,] 2.0874882 5.7249891
#>  [35,] 0.6284585 7.1840189
#>  [36,] 0.0000000 7.8124773
#>  [37,] 1.3225008 6.4899765
#>  [38,] 0.7815315 7.0309458
#>  [39,] 1.2896331 6.5228443
#>  [40,] 0.6302724 7.1822049
#>  [41,] 1.1712418 6.6412355
#>  [42,] 1.9527733 5.8597040
#>  [43,] 1.0176108 6.7948666
#>  [44,] 0.0000000 7.8124773
#>  [45,] 0.0000000 7.8124773
#>  [46,] 0.0000000 7.8124773
#>  [47,] 2.6996230 5.1128543
#>  [48,] 0.7486637 7.0638136
#>  [49,] 0.0000000 7.8124773
#>  [50,] 2.0455050 5.7669723
#>  [51,] 0.0000000 7.8124773
#>  [52,] 1.1694278 6.6430495
#>  [53,] 2.3401116 5.4723657
#>  [54,] 1.5301952 6.2822821
#>  [55,] 2.4294147 5.3830626
#>  [56,] 1.3789362 6.4335412
#>  [57,] 2.1802573 5.6322200
#>  [58,] 2.4517216 5.3607557
#>  [59,] 2.4517216 5.3607557
#>  [60,] 1.2075382 6.6049391
#>  [61,] 1.8359967 5.9764807
#>  [62,] 1.5880952 6.2243821
#>  [63,] 0.6665689 7.1459085
#>  [64,] 1.5880952 6.2243821
#>  [65,] 2.8690197 4.9434576
#>  [66,] 1.3789362 6.4335412
#>  [67,] 6.3972447 1.4152326
#>  [68,] 6.5605939 1.2518834
#>  [69,] 5.4980253 2.3144520
#>  [70,] 5.9746667 1.8378106
#>  [71,] 6.5156360 1.2968413
#>  [72,] 5.4980253 2.3144520
#>  [73,] 4.7280633 3.0844141
#>  [74,] 5.6534629 2.1590144
#>  [75,] 5.3467662 2.4657111
#>  [76,] 7.1459085 0.6665689
#>  [77,] 5.3946321 2.4178452
#>  [78,] 6.6465155 1.1659618
#>  [79,] 5.0747439 2.7377334
#>  [80,] 7.8124773 0.0000000
#>  [81,] 7.2715080 0.5409693
#>  [82,] 5.0578722 2.7546051
#>  [83,] 5.0965307 2.7159467
#>  [84,] 6.5228443 1.2896331
#>  [85,] 6.5228443 1.2896331
#>  [86,] 6.4024873 1.4099900
#>  [87,] 6.0162431 1.7962343
#>  [88,] 7.1822049 0.6302724
#>  [89,] 7.1822049 0.6302724
#>  [90,] 6.6412355 1.1712418
#>  [91,] 6.1664081 1.6460692
#>  [92,] 7.1840189 0.6284585
#>  [93,] 5.7722149 2.0402624
#>  [94,] 6.5537464 1.2587309
#>  [95,] 5.3467662 2.4657111
#>  [96,] 5.4159304 2.3965469
#>  [97,] 6.4335412 1.3789362
#>  [98,] 6.5156360 1.2968413
#>  [99,] 5.5052335 2.3072438
#> [100,] 6.7948666 1.0176108
#> [101,] 7.8124773 0.0000000
#> [102,] 6.8909509 0.9215264
#> [103,] 5.8940251 1.9184522
#> [104,] 5.1453614 2.6671160
#> [105,] 6.1800369 1.6324404
#> [106,] 5.5515784 2.2608989
#> [107,] 5.6390676 2.1734098
#> [108,] 6.4731048 1.3393725
#> [109,] 6.4731048 1.3393725
#> [110,] 6.6430495 1.1694278
#> [111,] 6.4731048 1.3393725
#> [112,] 5.6254388 2.1870385
#> [113,] 5.0303987 2.7820786
#> [114,] 6.5605939 1.2518834
#> [115,] 5.8943858 1.9180915
#> [116,] 4.5764989 3.2359784
#> [117,] 5.4173837 2.3950936
#> [118,] 5.3024210 2.5100563
#> [119,] 5.6915733 2.1209040
#> [120,] 5.5429832 2.2694941
#> [121,] 7.1015633 0.7109140
#> [122,] 5.8615180 1.9509593
#> [123,] 5.4173837 2.3950936
#> [124,] 6.2538972 1.5585801
#> [125,] 6.7948666 1.0176108
#> [126,] 6.4899765 1.3225008
#> [127,] 6.0133351 1.7991423
#> [128,] 6.7948666 1.0176108
#> [129,] 7.1840189 0.6284585
#> [130,] 6.0839525 1.7285248
#> [131,] 7.8124773 0.0000000
#> [132,] 7.0638136 0.7486637
#> [133,] 6.7948666 1.0176108
#> [134,] 7.8124773 0.0000000
#> [135,] 7.1015633 0.7109140
#> [136,] 7.2715080 0.5409693
#> [137,] 5.8047220 2.0077553
#> [138,] 6.2538972 1.5585801
#> [139,] 5.8119302 2.0005471
#> 
#> $prob
#>              [,1]       [,2]
#>   [1,] 0.32556187 0.67443813
#>   [2,] 0.29422787 0.70577213
#>   [3,] 0.09582923 0.90417077
#>   [4,] 0.27147650 0.72852350
#>   [5,] 0.27252106 0.72747894
#>   [6,] 0.09582923 0.90417077
#>   [7,] 0.19863083 0.80136917
#>   [8,] 0.26115435 0.73884565
#>   [9,] 0.23029088 0.76970912
#>  [10,] 0.27654066 0.72345934
#>  [11,] 0.24068445 0.75931555
#>  [12,] 0.00000000 1.00000000
#>  [13,] 0.00000000 1.00000000
#>  [14,] 0.00000000 1.00000000
#>  [15,] 0.30478221 0.69521779
#>  [16,] 0.31523971 0.68476029
#>  [17,] 0.38448399 0.61551601
#>  [18,] 0.08044292 0.91955708
#>  [19,] 0.09582923 0.90417077
#>  [20,] 0.14968719 0.85031281
#>  [21,] 0.23029088 0.76970912
#>  [22,] 0.17650434 0.82349566
#>  [23,] 0.13025455 0.86974545
#>  [24,] 0.00000000 1.00000000
#>  [25,] 0.26764291 0.73235709
#>  [26,] 0.19103358 0.80896642
#>  [27,] 0.26603249 0.73396751
#>  [28,] 0.34787437 0.65212563
#>  [29,] 0.13025455 0.86974545
#>  [30,] 0.00000000 1.00000000
#>  [31,] 0.24024081 0.75975919
#>  [32,] 0.27147650 0.72852350
#>  [33,] 0.13025455 0.86974545
#>  [34,] 0.26719927 0.73280073
#>  [35,] 0.08044292 0.91955708
#>  [36,] 0.00000000 1.00000000
#>  [37,] 0.16928060 0.83071940
#>  [38,] 0.10003632 0.89996368
#>  [39,] 0.16507351 0.83492649
#>  [40,] 0.08067511 0.91932489
#>  [41,] 0.14991938 0.85008062
#>  [42,] 0.24995571 0.75004429
#>  [43,] 0.13025455 0.86974545
#>  [44,] 0.00000000 1.00000000
#>  [45,] 0.00000000 1.00000000
#>  [46,] 0.00000000 1.00000000
#>  [47,] 0.34555275 0.65444725
#>  [48,] 0.09582923 0.90417077
#>  [49,] 0.00000000 1.00000000
#>  [50,] 0.26182540 0.73817460
#>  [51,] 0.00000000 1.00000000
#>  [52,] 0.14968719 0.85031281
#>  [53,] 0.29953515 0.70046485
#>  [54,] 0.19586556 0.80413444
#>  [55,] 0.31096598 0.68903402
#>  [56,] 0.17650434 0.82349566
#>  [57,] 0.27907374 0.72092626
#>  [58,] 0.31382128 0.68617872
#>  [59,] 0.31382128 0.68617872
#>  [60,] 0.15456534 0.84543466
#>  [61,] 0.23500825 0.76499175
#>  [62,] 0.20327678 0.79672322
#>  [63,] 0.08532106 0.91467894
#>  [64,] 0.20327678 0.79672322
#>  [65,] 0.36723559 0.63276441
#>  [66,] 0.17650434 0.82349566
#>  [67,] 0.81884970 0.18115030
#>  [68,] 0.83975846 0.16024154
#>  [69,] 0.70374928 0.29625072
#>  [70,] 0.76475956 0.23524044
#>  [71,] 0.83400383 0.16599617
#>  [72,] 0.70374928 0.29625072
#>  [73,] 0.60519385 0.39480615
#>  [74,] 0.72364535 0.27635465
#>  [75,] 0.68438806 0.31561194
#>  [76,] 0.91467894 0.08532106
#>  [77,] 0.69051492 0.30948508
#>  [78,] 0.85075645 0.14924355
#>  [79,] 0.64956910 0.35043090
#>  [80,] 1.00000000 0.00000000
#>  [81,] 0.93075572 0.06924428
#>  [82,] 0.64740953 0.35259047
#>  [83,] 0.65235782 0.34764218
#>  [84,] 0.83492649 0.16507351
#>  [85,] 0.83492649 0.16507351
#>  [86,] 0.81952076 0.18047924
#>  [87,] 0.77008135 0.22991865
#>  [88,] 0.91932489 0.08067511
#>  [89,] 0.91932489 0.08067511
#>  [90,] 0.85008062 0.14991938
#>  [91,] 0.78930253 0.21069747
#>  [92,] 0.91955708 0.08044292
#>  [93,] 0.73884565 0.26115435
#>  [94,] 0.83888198 0.16111802
#>  [95,] 0.68438806 0.31561194
#>  [96,] 0.69324110 0.30675890
#>  [97,] 0.82349566 0.17650434
#>  [98,] 0.83400383 0.16599617
#>  [99,] 0.70467193 0.29532807
#> [100,] 0.86974545 0.13025455
#> [101,] 1.00000000 0.00000000
#> [102,] 0.88204428 0.11795572
#> [103,] 0.75443740 0.24556260
#> [104,] 0.65860817 0.34139183
#> [105,] 0.79104702 0.20895298
#> [106,] 0.71060410 0.28939590
#> [107,] 0.72180274 0.27819726
#> [108,] 0.82855982 0.17144018
#> [109,] 0.82855982 0.17144018
#> [110,] 0.85031281 0.14968719
#> [111,] 0.82855982 0.17144018
#> [112,] 0.72005825 0.27994175
#> [113,] 0.64389290 0.35610710
#> [114,] 0.83975846 0.16024154
#> [115,] 0.75448357 0.24551643
#> [116,] 0.58579356 0.41420644
#> [117,] 0.69342712 0.30657288
#> [118,] 0.67871186 0.32128814
#> [119,] 0.72852350 0.27147650
#> [120,] 0.70950391 0.29049609
#> [121,] 0.90900274 0.09099726
#> [122,] 0.75027648 0.24972352
#> [123,] 0.69342712 0.30657288
#> [124,] 0.80050117 0.19949883
#> [125,] 0.86974545 0.13025455
#> [126,] 0.83071940 0.16928060
#> [127,] 0.76970912 0.23029088
#> [128,] 0.86974545 0.13025455
#> [129,] 0.91955708 0.08044292
#> [130,] 0.77874819 0.22125181
#> [131,] 1.00000000 0.00000000
#> [132,] 0.90417077 0.09582923
#> [133,] 0.86974545 0.13025455
#> [134,] 1.00000000 0.00000000
#> [135,] 0.90900274 0.09099726
#> [136,] 0.93075572 0.06924428
#> [137,] 0.74300657 0.25699343
#> [138,] 0.80050117 0.19949883
#> [139,] 0.74392923 0.25607077
#> 
#> $class
#>   [1] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [19] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [37] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R"
#>  [55] "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "R" "M" "M" "M" "M" "M" "M"
#>  [73] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#>  [91] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [109] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> [127] "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M" "M"
#> 
#> $importance
#>         V1        V10        V11        V12        V13        V14        V15 
#>  1.2147225  2.6935064 14.3783674  2.8369587  0.0000000  0.0000000  0.0000000 
#>        V16        V17        V18        V19         V2        V20        V21 
#>  6.0690046  0.0000000  1.2185956  0.0000000  0.0000000  1.8302163  5.4333260 
#>        V22        V23        V24        V25        V26        V27        V28 
#>  1.6630249  6.5603406  0.0000000  0.0000000  0.0000000  1.7047314  1.7407731 
#>        V29         V3        V30        V31        V32        V33        V34 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V35        V36        V37        V38        V39         V4        V40 
#>  0.0000000  2.6943131  4.3181100  0.0000000  0.0000000  2.7252775  0.0000000 
#>        V41        V42        V43        V44        V45        V46        V47 
#>  3.7274478  0.0000000  0.0000000  0.0000000  8.2388159  4.8445424  0.0000000 
#>        V48        V49         V5        V50        V51        V52        V53 
#>  2.0750279  0.0000000  0.0000000  1.7414617  2.5892499  0.0000000  0.0000000 
#>        V54        V55        V56        V57        V58        V59         V6 
#>  3.7877722  3.4387415  1.2310198  0.0000000  0.4731733  0.0000000  0.0000000 
#>        V60         V7         V8         V9 
#>  0.8363534  0.0000000  1.5461114  8.3890148 
#> 
#> $terms
#> Class ~ V1 + V10 + V11 + V12 + V13 + V14 + V15 + V16 + V17 + 
#>     V18 + V19 + V2 + V20 + V21 + V22 + V23 + V24 + V25 + V26 + 
#>     V27 + V28 + V29 + V3 + V30 + V31 + V32 + V33 + V34 + V35 + 
#>     V36 + V37 + V38 + V39 + V4 + V40 + V41 + V42 + V43 + V44 + 
#>     V45 + V46 + V47 + V48 + V49 + V5 + V50 + V51 + V52 + V53 + 
#>     V54 + V55 + V56 + V57 + V58 + V59 + V6 + V60 + V7 + V8 + 
#>     V9
#> attr(,"variables")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"factors")
#>       V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26
#> Class  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V1     1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V10    0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V11    0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V12    0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V13    0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V14    0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V15    0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0   0   0   0
#> V16    0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0   0   0   0
#> V17    0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0   0   0   0
#> V18    0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0   0   0   0
#> V19    0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0   0   0   0
#> V2     0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0   0   0   0
#> V20    0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0   0   0   0
#> V21    0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0   0   0   0
#> V22    0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0   0   0   0
#> V23    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1   0   0   0
#> V24    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   1   0   0
#> V25    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   1   0
#> V26    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   1
#> V27    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V28    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V29    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V3     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V30    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V31    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V32    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V33    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V34    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V35    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V36    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V37    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V38    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V39    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V4     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V40    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V41    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V42    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V43    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V44    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V45    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V46    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V47    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V48    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V49    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V5     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V50    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V51    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V52    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V53    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V54    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V55    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V56    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V57    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V58    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V59    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V6     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V60    0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V7     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V8     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#> V9     0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0   0   0   0
#>       V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43
#> Class   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V1      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V10     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V11     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V12     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V13     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V14     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V15     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V16     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V17     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V18     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V19     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V2      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V20     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V21     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V22     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V23     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V24     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V25     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V26     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V27     1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V28     0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V29     0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V3      0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V30     0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V31     0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V32     0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0   0   0   0
#> V33     0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0   0   0   0
#> V34     0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0   0   0   0
#> V35     0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0   0   0   0
#> V36     0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0   0   0   0
#> V37     0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0   0   0   0
#> V38     0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0   0   0   0
#> V39     0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0   0   0   0
#> V4      0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0   0   0   0
#> V40     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1   0   0   0
#> V41     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   1   0   0
#> V42     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   1   0
#> V43     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   1
#> V44     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V45     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V46     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V47     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V48     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V49     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V5      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V50     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V51     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V52     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V53     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V54     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V55     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V56     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V57     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V58     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V59     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V6      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V60     0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V7      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V8      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#> V9      0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0   0   0   0
#>       V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60
#> Class   0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V1      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V10     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V11     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V12     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V13     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V14     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V15     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V16     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V17     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V18     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V19     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V2      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V20     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V21     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V22     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V23     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V24     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V25     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V26     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V27     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V28     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V29     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V3      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V30     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V31     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V32     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V33     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V34     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V35     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V36     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V37     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V38     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V39     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V4      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V40     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V41     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V42     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V43     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V44     1   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V45     0   1   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V46     0   0   1   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V47     0   0   0   1   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V48     0   0   0   0   1   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V49     0   0   0   0   0   1  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V5      0   0   0   0   0   0  1   0   0   0   0   0   0   0   0   0   0  0   0
#> V50     0   0   0   0   0   0  0   1   0   0   0   0   0   0   0   0   0  0   0
#> V51     0   0   0   0   0   0  0   0   1   0   0   0   0   0   0   0   0  0   0
#> V52     0   0   0   0   0   0  0   0   0   1   0   0   0   0   0   0   0  0   0
#> V53     0   0   0   0   0   0  0   0   0   0   1   0   0   0   0   0   0  0   0
#> V54     0   0   0   0   0   0  0   0   0   0   0   1   0   0   0   0   0  0   0
#> V55     0   0   0   0   0   0  0   0   0   0   0   0   1   0   0   0   0  0   0
#> V56     0   0   0   0   0   0  0   0   0   0   0   0   0   1   0   0   0  0   0
#> V57     0   0   0   0   0   0  0   0   0   0   0   0   0   0   1   0   0  0   0
#> V58     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   1   0  0   0
#> V59     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   1  0   0
#> V6      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  1   0
#> V60     0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   1
#> V7      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V8      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#> V9      0   0   0   0   0   0  0   0   0   0   0   0   0   0   0   0   0  0   0
#>       V7 V8 V9
#> Class  0  0  0
#> V1     0  0  0
#> V10    0  0  0
#> V11    0  0  0
#> V12    0  0  0
#> V13    0  0  0
#> V14    0  0  0
#> V15    0  0  0
#> V16    0  0  0
#> V17    0  0  0
#> V18    0  0  0
#> V19    0  0  0
#> V2     0  0  0
#> V20    0  0  0
#> V21    0  0  0
#> V22    0  0  0
#> V23    0  0  0
#> V24    0  0  0
#> V25    0  0  0
#> V26    0  0  0
#> V27    0  0  0
#> V28    0  0  0
#> V29    0  0  0
#> V3     0  0  0
#> V30    0  0  0
#> V31    0  0  0
#> V32    0  0  0
#> V33    0  0  0
#> V34    0  0  0
#> V35    0  0  0
#> V36    0  0  0
#> V37    0  0  0
#> V38    0  0  0
#> V39    0  0  0
#> V4     0  0  0
#> V40    0  0  0
#> V41    0  0  0
#> V42    0  0  0
#> V43    0  0  0
#> V44    0  0  0
#> V45    0  0  0
#> V46    0  0  0
#> V47    0  0  0
#> V48    0  0  0
#> V49    0  0  0
#> V5     0  0  0
#> V50    0  0  0
#> V51    0  0  0
#> V52    0  0  0
#> V53    0  0  0
#> V54    0  0  0
#> V55    0  0  0
#> V56    0  0  0
#> V57    0  0  0
#> V58    0  0  0
#> V59    0  0  0
#> V6     0  0  0
#> V60    0  0  0
#> V7     1  0  0
#> V8     0  1  0
#> V9     0  0  1
#> attr(,"term.labels")
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,"predvars")
#> list(Class, V1, V10, V11, V12, V13, V14, V15, V16, V17, V18, 
#>     V19, V2, V20, V21, V22, V23, V24, V25, V26, V27, V28, V29, 
#>     V3, V30, V31, V32, V33, V34, V35, V36, V37, V38, V39, V4, 
#>     V40, V41, V42, V43, V44, V45, V46, V47, V48, V49, V5, V50, 
#>     V51, V52, V53, V54, V55, V56, V57, V58, V59, V6, V60, V7, 
#>     V8, V9)
#> attr(,"dataClasses")
#>     Class        V1       V10       V11       V12       V13       V14       V15 
#>  "factor" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V16       V17       V18       V19        V2       V20       V21       V22 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V23       V24       V25       V26       V27       V28       V29        V3 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V30       V31       V32       V33       V34       V35       V36       V37 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V38       V39        V4       V40       V41       V42       V43       V44 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V45       V46       V47       V48       V49        V5       V50       V51 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>       V52       V53       V54       V55       V56       V57       V58       V59 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        V6       V60        V7        V8        V9 
#> "numeric" "numeric" "numeric" "numeric" "numeric" 
#> 
#> $call
#> adabag::boosting(formula = formula, data = data, mfinal = 10L, 
#>     control = list(minsplit = 20L, minbucket = 7, cp = 0.01, 
#>         maxcompete = 4L, maxsurrogate = 5L, usesurrogate = 2L, 
#>         surrogatestyle = 0L, maxdepth = 30L, xval = 0L))
#> 
#> attr(,"vardep.summary")
#>  M  R 
#> 73 66 
#> attr(,"class")
#> [1] "boosting"
print(learner$importance())
#>        V11         V9        V45        V23        V16        V21        V46 
#> 14.3783674  8.3890148  8.2388159  6.5603406  6.0690046  5.4333260  4.8445424 
#>        V37        V54        V41        V55        V12         V4        V36 
#>  4.3181100  3.7877722  3.7274478  3.4387415  2.8369587  2.7252775  2.6943131 
#>        V10        V51        V48        V20        V50        V28        V27 
#>  2.6935064  2.5892499  2.0750279  1.8302163  1.7414617  1.7407731  1.7047314 
#>        V22         V8        V56        V18         V1        V60        V58 
#>  1.6630249  1.5461114  1.2310198  1.2185956  1.2147225  0.8363534  0.4731733 
#>        V13        V14        V15        V17        V19         V2        V24 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V25        V26        V29         V3        V30        V31        V32 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V33        V34        V35        V38        V39        V40        V42 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V43        V44        V47        V49         V5        V52        V53 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>        V57        V59         V6         V7 
#>  0.0000000  0.0000000  0.0000000  0.0000000 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551