BlockForest Classification Learner

Random forests for blocks of clinical and omics covariate data. Calls blockForest::blockfor() from package blockForest.

In this learner, only the trained forest object ($forest) is retained. The optimized block-specific tuning parameters (paramvalues) and the biased OOB error estimate (biased_oob_error_donotuse) are discarded, as they are either not needed for downstream use or not reliable for performance estimation.

Initial parameter values

num.threads is initialized to 1 to avoid conflicts with parallelization via future.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.blockforest")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, blockForest

Parameters

Id	Type	Default	Levels	Range

blocks	untyped	-		-

block.method	character	BlockForest	BlockForest, RandomBlock, BlockVarSel, VarProb, SplitWeights	-

num.trees	integer	2000		\([1, \infty)\)

mtry	untyped	NULL		-

nsets	integer	300		\([1, \infty)\)

num.trees.pre	integer	1500		\([1, \infty)\)

splitrule	character	extratrees	extratrees, gini	-

always.select.block	integer	0		\([0, 1]\)

importance	character	-	none, impurity, impurity_corrected, permutation	-

num.threads	integer	-		\([1, \infty)\)

seed	integer	NULL		\((-\infty, \infty)\)

verbose	logical	TRUE	TRUE, FALSE	-

References

Hornung, R., Wright, N. M (2019). “Block Forests: Random forests for blocks of clinical and omics covariate data.” BMC Bioinformatics, 20(1), 1–17. doi:10.1186/s12859-019-2942-y , https://doi.org/10.1186/s12859-019-2942-y.

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifBlockForest

Methods

Public methods

LearnerClassifBlockForest$new()

LearnerClassifBlockForest$importance()

LearnerClassifBlockForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifBlockForest$new()

Method importance()

The importance scores are extracted from the model slot variable.importance.

Usage

LearnerClassifBlockForest$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifBlockForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define a Task
task = tsk("sonar")
# Create train and test set
ids = partition(task)
# check task's features
task$feature_names
#>  [1] "V1"  "V10" "V11" "V12" "V13" "V14" "V15" "V16" "V17" "V18" "V19" "V2" 
#> [13] "V20" "V21" "V22" "V23" "V24" "V25" "V26" "V27" "V28" "V29" "V3"  "V30"
#> [25] "V31" "V32" "V33" "V34" "V35" "V36" "V37" "V38" "V39" "V4"  "V40" "V41"
#> [37] "V42" "V43" "V44" "V45" "V46" "V47" "V48" "V49" "V5"  "V50" "V51" "V52"
#> [49] "V53" "V54" "V55" "V56" "V57" "V58" "V59" "V6"  "V60" "V7"  "V8"  "V9" 
# partition features to 2 blocks
blocks = list(bl1 = 1:42, bl2 = 43:60)
# define learner
learner = lrn("classif.blockforest", blocks = blocks,
              importance = "permutation", nsets = 10, predict_type = "prob",
              num.trees = 50, num.trees.pre = 10, splitrule = "gini")
# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
# feature importance
learner$importance()
#>           V11           V10            V9           V12           V49 
#>  2.629191e-02  1.721961e-02  1.673982e-02  1.376528e-02  1.292431e-02 
#>           V48           V36           V13           V17           V47 
#>  1.175342e-02  1.166384e-02  1.063086e-02  9.181122e-03  5.711703e-03 
#>           V22           V45           V19           V43            V5 
#>  4.780953e-03  4.700673e-03  4.558102e-03  4.230199e-03  3.992506e-03 
#>           V23           V20           V21            V1           V33 
#>  3.973354e-03  3.691129e-03  3.689008e-03  3.166448e-03  3.061601e-03 
#>           V29           V15           V27           V18           V28 
#>  2.947345e-03  2.937167e-03  2.917970e-03  2.808775e-03  2.792004e-03 
#>           V56            V8            V4           V52           V25 
#>  2.759589e-03  2.429142e-03  2.385770e-03  2.385559e-03  2.376640e-03 
#>           V46            V2           V14           V37           V39 
#>  2.278584e-03  2.208459e-03  2.027480e-03  1.930593e-03  1.724938e-03 
#>           V44           V30           V60           V50           V58 
#>  1.676819e-03  1.263602e-03  1.157216e-03  9.641633e-04  8.753668e-04 
#>           V16           V31           V59           V32           V53 
#>  7.654001e-04  5.854748e-04  5.710979e-04  3.601652e-04  3.548226e-04 
#>           V40           V34           V26           V55           V41 
#>  2.637486e-04  2.297439e-04  2.259026e-04  1.208673e-04 -2.722161e-06 
#>           V51           V57           V35            V7            V6 
#> -2.135775e-05 -5.275054e-05 -2.163265e-04 -4.250000e-04 -4.584396e-04 
#>           V54           V38            V3           V24           V42 
#> -7.908556e-04 -9.230003e-04 -1.664561e-03 -1.698186e-03 -1.760691e-03 
# Make predictions for the test observations
pred = learner$predict(task, row_ids = ids$test)
pred
#> 
#> ── <PredictionClassif> for 69 observations: ────────────────────────────────────
#>  row_ids truth response    prob.M     prob.R
#>        1     R        M 0.5090000 0.49100000
#>        7     R        M 0.7072778 0.29272222
#>        8     R        M 0.7166667 0.28333333
#>      ---   ---      ---       ---        ---
#>      194     M        M 0.6672222 0.33277778
#>      195     M        M 0.9092857 0.09071429
#>      206     M        M 0.8558333 0.14416667
# Score the predictions
pred$score()
#> classif.ce 
#>  0.3623188