Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.2138268 -0.09941681  0.02082315 -0.02645629 -0.004082530 -0.01885153
#>  [2,] -0.3097231 -0.12754559 -0.11702363 -0.03730854 -0.021209250 -0.07192567
#>  [3,] -0.3266691 -0.09571510 -0.05087063 -0.03046938 -0.066036841 -0.07180578
#>  [4,] -0.3272073 -0.17425906 -0.06314577 -0.14127945 -0.129357146 -0.05441376
#>  [5,] -0.3505762 -0.17490530 -0.10525030 -0.07035825 -0.174329699 -0.08336323
#>  [6,] -0.2755682 -0.10925003 -0.13652303 -0.12889582 -0.162494264 -0.17559862
#>  [7,] -0.3983407 -0.12109460 -0.03150091 -0.16004847 -0.151962146 -0.01873699
#>  [8,] -0.1557077 -0.12246935 -0.21393436 -0.06021576  0.004505543 -0.03128998
#>  [9,] -0.3755364 -0.16209103 -0.19676714 -0.05394801 -0.109772048 -0.01679979
#> [10,] -0.2573214 -0.20468830 -0.17045986 -0.08522316  0.001286916  0.02488880
#>                V15          V16         V17          V18          V19
#>  [1,]  0.022063606 -0.036407522  0.07600709  0.101831838  0.069651515
#>  [2,] -0.034792683 -0.091787182  0.03740484 -0.006971980 -0.036009234
#>  [3,]  0.002792816 -0.033377020  0.09881313  0.070897241  0.070127375
#>  [4,] -0.015460377  0.026264974 -0.00404679  0.058683238  0.060623736
#>  [5,] -0.035408166 -0.103068844 -0.07057463 -0.068014181  0.118693267
#>  [6,] -0.065861674  0.095358076 -0.10320382  0.026641106  0.008280649
#>  [7,] -0.006863313  0.025484349  0.03441342  0.057635067  0.113452320
#>  [8,] -0.075144323 -0.008924608 -0.15676119 -0.199113431 -0.127994962
#>  [9,] -0.102632340  0.006871366  0.06570053 -0.004534642 -0.024085192
#> [10,]  0.018785482  0.029568048  0.03660604 -0.038222219  0.134333728
#>               V2         V20           V21         V22         V23         V24
#>  [1,] -0.3236364 0.015672904  0.1689857775  0.21099264  0.21434233  0.23218333
#>  [2,] -0.3655441 0.004765308  0.0129359781  0.10067988 -0.01422315  0.03523984
#>  [3,] -0.3242825 0.088933265  0.1747999671  0.26023717  0.13836058  0.22775366
#>  [4,] -0.2156351 0.114146849 -0.0099964666  0.06196059  0.12683946  0.04862742
#>  [5,] -0.2872457 0.115180117  0.0001138374  0.12402375  0.03249875  0.10675794
#>  [6,] -0.3494902 0.058944355  0.0017976663  0.09364412  0.15323578  0.08916160
#>  [7,] -0.3109724 0.066819907  0.1215972711  0.08399432  0.13184022  0.17559549
#>  [8,] -0.2514669 0.025821887 -0.1492411537 -0.07012267 -0.03189531 -0.04816872
#>  [9,] -0.2868105 0.005528360  0.0394393554  0.13988913  0.05802262  0.23504010
#> [10,] -0.3649336 0.092319989  0.1894658698  0.11462146  0.12763997  0.19138281
#>                V25          V26        V27        V28         V29         V3
#>  [1,]  0.146087125  0.138187312 0.12233663 0.11547664  0.07318180 -0.3108992
#>  [2,]  0.036824199  0.197588283 0.16493662 0.24989162  0.20659864 -0.3579591
#>  [3,]  0.202534918  0.239106588 0.12766489 0.29029229  0.17384274 -0.2929955
#>  [4,]  0.146917983  0.027690487 0.11962053 0.01435082  0.04760314 -0.2818501
#>  [5,]  0.173470763 -0.005735569 0.05166105 0.02917678 -0.07783385 -0.2169170
#>  [6,]  0.170814109  0.054584456 0.16891558 0.10204093  0.08704971 -0.2404479
#>  [7,]  0.102054109  0.095731064 0.06738560 0.06093651  0.16613871 -0.2466311
#>  [8,] -0.009041818  0.026367316 0.04966178 0.01445188 -0.03575059 -0.1654873
#>  [9,]  0.114323877  0.180878807 0.16121113 0.06385580  0.09088376 -0.2564889
#> [10,]  0.182634262  0.208545817 0.09872570 0.17280798  0.08677488 -0.3440002
#>              V30         V31         V32          V33          V34         V35
#>  [1,] 0.05476268  0.03750404 -0.08440954 -0.086557467 -0.036212771 -0.07263951
#>  [2,] 0.10270654  0.04223481  0.11384680  0.080823609  0.094737865  0.04617165
#>  [3,] 0.10955484  0.11761504  0.03980305  0.029980575 -0.052359518 -0.05381666
#>  [4,] 0.05143818 -0.04240392 -0.07806551 -0.173680923 -0.006904248 -0.13998802
#>  [5,] 0.07603828 -0.12148641 -0.06700528 -0.001590739 -0.009956459 -0.08671230
#>  [6,] 0.08844493  0.07477442  0.01178562 -0.018401254 -0.069570402 -0.14610171
#>  [7,] 0.12319095  0.01571896 -0.07823731 -0.053234362 -0.120358048 -0.09270644
#>  [8,] 0.03673350 -0.06022433 -0.10750237 -0.011675092 -0.001957751 -0.07154975
#>  [9,] 0.10167070  0.05470769 -0.05271044  0.029611575  0.044257318  0.07319530
#> [10,] 0.10430523  0.02998943  0.10134766  0.049329124 -0.011603202 -0.03233713
#>               V36          V37          V38         V39         V4         V40
#>  [1,]  0.01862584 -0.012145883  0.018294425 -0.02988729 -0.2452676 -0.10115306
#>  [2,] -0.07255513 -0.019850243  0.005764404 -0.12360992 -0.2950275 -0.03019155
#>  [3,] -0.01674580  0.026282376 -0.005987378  0.03123386 -0.2753836 -0.09316102
#>  [4,] -0.06248227 -0.056899286 -0.131638447 -0.05605059 -0.1881832 -0.07765785
#>  [5,] -0.11524316 -0.137180417 -0.125037429 -0.09232101 -0.2845071 -0.03780640
#>  [6,] -0.09488368 -0.037562440 -0.111034252 -0.17558594 -0.3075855 -0.07853758
#>  [7,] -0.02514031 -0.115328624 -0.171083044 -0.14701449 -0.3207503 -0.05305490
#>  [8,] -0.16389202 -0.065697986 -0.127109260 -0.04917503 -0.2737403 -0.11720979
#>  [9,] -0.05945358 -0.044101896  0.055211010  0.01857666 -0.1856399  0.01535951
#> [10,] -0.06538003  0.008526885 -0.112518397 -0.10900174 -0.3532054 -0.03778561
#>               V41          V42          V43          V44         V45
#>  [1,] -0.08926096 -0.065039359 -0.074972085 -0.095452525 -0.14972241
#>  [2,] -0.01875301  0.033985947 -0.147486253 -0.015206370 -0.04334841
#>  [3,] -0.03128251 -0.111421798 -0.148937442 -0.075016834 -0.08499428
#>  [4,] -0.18572659 -0.212645462 -0.136315270 -0.061713883 -0.16637491
#>  [5,] -0.10094743 -0.051258892 -0.129595626 -0.226768743 -0.20574928
#>  [6,] -0.14143383 -0.091449554 -0.082376199 -0.151544430 -0.13763115
#>  [7,] -0.10738444 -0.073444348 -0.124526729 -0.195211156 -0.14120824
#>  [8,] -0.01669596 -0.081286902 -0.072969939 -0.055106033 -0.06430160
#>  [9,]  0.01664592 -0.001895231  0.005590535 -0.148167222 -0.06626211
#> [10,]  0.06705184  0.018178334  0.038743340 -0.004304292 -0.13852953
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.13034111 -0.2002546 -0.1968730 -0.2309081 -0.1780304 -0.3343658
#>  [2,] -0.19354006 -0.1815937 -0.2687372 -0.2420183 -0.2743830 -0.2780791
#>  [3,] -0.08802227 -0.1697980 -0.2435283 -0.3310932 -0.3281962 -0.3524907
#>  [4,] -0.23406431 -0.2692286 -0.1323580 -0.3491465 -0.2647510 -0.3712617
#>  [5,] -0.14936193 -0.2593703 -0.2749147 -0.1430813 -0.1948266 -0.2374404
#>  [6,] -0.11778726 -0.2153606 -0.2425864 -0.1861195 -0.1830005 -0.2782753
#>  [7,] -0.15437907 -0.1037729 -0.2425978 -0.3719089 -0.3509283 -0.4522211
#>  [8,] -0.18757340 -0.1818517 -0.1742458 -0.2282040 -0.1319990 -0.1444924
#>  [9,] -0.16322065 -0.1700321 -0.1237969 -0.3502414 -0.2051972 -0.2368814
#> [10,] -0.03204466 -0.2454711 -0.2475810 -0.3279341 -0.1492187 -0.3873537
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2515218 -0.3404049 -0.3299901 -0.3786593 -0.3076276 -0.3930788
#>  [2,] -0.3234737 -0.4056740 -0.3281478 -0.4053428 -0.3958175 -0.3596335
#>  [3,] -0.3463285 -0.3874664 -0.2623062 -0.4455403 -0.3043486 -0.3483806
#>  [4,] -0.3903158 -0.3261872 -0.4094126 -0.2764187 -0.2263990 -0.4134371
#>  [5,] -0.3640737 -0.2161198 -0.3155899 -0.2164101 -0.2397477 -0.3946756
#>  [6,] -0.3351004 -0.2340710 -0.3173681 -0.3547050 -0.3118007 -0.2929380
#>  [7,] -0.4098100 -0.3924772 -0.4283415 -0.3985092 -0.4690581 -0.3224071
#>  [8,] -0.1753603 -0.1813135 -0.1717236 -0.3274600 -0.2597184 -0.3207338
#>  [9,] -0.2452829 -0.3313183 -0.2477088 -0.3185117 -0.3382768 -0.3063177
#> [10,] -0.2673339 -0.3549499 -0.3820461 -0.3214822 -0.4156089 -0.3151517
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.2929751 -0.4002023 -0.4246099 -0.1468524 -0.3075418 -0.1039100
#>  [2,] -0.2936351 -0.3104834 -0.3309010 -0.1982352 -0.3234499 -0.1954352
#>  [3,] -0.3318888 -0.4159880 -0.3260267 -0.2533813 -0.3105536 -0.1813971
#>  [4,] -0.3047084 -0.2299309 -0.3695961 -0.1864643 -0.3923986 -0.1253685
#>  [5,] -0.2460689 -0.3334968 -0.2399896 -0.1475748 -0.4092196 -0.1023056
#>  [6,] -0.3988505 -0.3734681 -0.3332823 -0.2955675 -0.2809627 -0.2641748
#>  [7,] -0.3659831 -0.3863192 -0.3274344 -0.2458150 -0.3388160 -0.2737588
#>  [8,] -0.2762858 -0.1939808 -0.1839925 -0.2675465 -0.1840949 -0.1669451
#>  [9,] -0.4075522 -0.3662175 -0.2598179 -0.1979368 -0.3465389 -0.2563318
#> [10,] -0.3866497 -0.4320720 -0.3935786 -0.2705621 -0.4764964 -0.1883890
#>                V8          V9
#>  [1,] -0.04990528 -0.09481193
#>  [2,] -0.17111328 -0.15149836
#>  [3,] -0.15833945 -0.11186215
#>  [4,] -0.10448819 -0.24346214
#>  [5,] -0.25860650 -0.24004569
#>  [6,] -0.20357462 -0.14088002
#>  [7,] -0.21239537 -0.16461005
#>  [8,] -0.17130736 -0.24426048
#>  [9,] -0.23614228 -0.08369194
#> [10,] -0.22222880 -0.14648701
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.02472326 -0.02604834  0.01841642 -0.08851775 -0.01627867 -0.05185060
#> [2,] -0.03587958  0.07435418 -0.05908088 -0.05743376  0.09029820  0.08329904
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.00160775 0.037138512 -0.01620926  0.02341774
#> [2,] -0.04516472 0.009646717 -0.01157124 -0.12022502
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -2.283798e-05 -2.005936e-04 -0.0002257829 -0.0001950675 -0.0001868184
#>  [2,]  1.480040e-04  1.140944e-03  0.0014582826  0.0015420048  0.0014737166
#>  [3,] -3.098196e-05 -2.498959e-04 -0.0002828289 -0.0002451788 -0.0002460066
#>  [4,]  2.277537e-05  2.008926e-04  0.0002650801  0.0002881776  0.0002716425
#>  [5,]  1.133412e-04  1.001089e-03  0.0013145763  0.0014383383  0.0013155302
#>  [6,]  1.573306e-04  1.335238e-03  0.0017562107  0.0019213609  0.0017684600
#>  [7,] -5.459871e-05 -4.579166e-04 -0.0005804021 -0.0006097766 -0.0005574593
#>  [8,] -4.243645e-05 -3.763344e-04 -0.0005051596 -0.0005666516 -0.0004867339
#>  [9,]  9.674610e-06  6.927947e-05  0.0000985231  0.0001148184  0.0001156349
#> [10,] -1.241006e-04 -9.763012e-04 -0.0011856269 -0.0011720490 -0.0011247116
#>                 V14           V15           V16           V17           V18
#>  [1,] -0.0001276731 -1.249221e-04 -6.971531e-05 -7.878835e-05 -7.929943e-05
#>  [2,]  0.0012233308  9.671671e-04  6.926209e-04  6.544637e-04  7.937647e-04
#>  [3,] -0.0001970947 -1.933098e-04 -1.312910e-04 -1.318742e-04 -1.305942e-04
#>  [4,]  0.0002153706  1.598903e-04  1.237804e-04  1.214274e-04  1.531732e-04
#>  [5,]  0.0009854230  7.017705e-04  4.439826e-04  4.211385e-04  5.806686e-04
#>  [6,]  0.0013396115  9.620803e-04  6.137225e-04  5.832738e-04  7.775647e-04
#>  [7,] -0.0004075304 -3.093509e-04 -1.712124e-04 -1.646501e-04 -2.127259e-04
#>  [8,] -0.0003299365 -2.006885e-04 -7.656245e-05 -5.483801e-05 -1.070210e-04
#>  [9,]  0.0001106416  8.066212e-05  8.115517e-05  7.807978e-05  9.126173e-05
#> [10,] -0.0009074886 -7.682764e-04 -5.258352e-04 -4.987936e-04 -5.622752e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0001583348 -2.103495e-05 -0.0003055203 -0.0002647129 -0.0001058082
#>  [2,]  0.0014957083  1.708149e-04  0.0021033984  0.0021917339  0.0018116422
#>  [3,] -0.0002338495 -3.077539e-05 -0.0004062382 -0.0003774668 -0.0002080609
#>  [4,]  0.0002855034  2.721751e-05  0.0003821393  0.0003982924  0.0003406154
#>  [5,]  0.0012356221  1.368440e-04  0.0017215147  0.0017436193  0.0013645869
#>  [6,]  0.0016656793  1.876044e-04  0.0023351993  0.0023817597  0.0018983994
#>  [7,] -0.0004860937 -6.132006e-05 -0.0007454106 -0.0007284127 -0.0004984939
#>  [8,] -0.0003659665 -5.551103e-05 -0.0005397410 -0.0005309583 -0.0003510757
#>  [9,]  0.0001504789  1.355536e-05  0.0001747015  0.0002100114  0.0002275468
#> [10,] -0.0010883616 -1.379041e-04 -0.0016810143 -0.0016833174 -0.0012054275
#>                 V23           V24           V25           V26           V27
#>  [1,]  8.512777e-05  1.791838e-04  1.526794e-04  1.666980e-04  1.866424e-04
#>  [2,]  1.169325e-03  6.519082e-04  2.560952e-04  3.909012e-04  7.462739e-04
#>  [3,] -7.844187e-06  9.863137e-05  1.008825e-04  9.599803e-05  9.996928e-05
#>  [4,]  2.246210e-04  1.373397e-04  7.017945e-05  8.668135e-05  1.507407e-04
#>  [5,]  6.743986e-04  1.434971e-04 -2.062765e-04 -1.488029e-04  2.228593e-04
#>  [6,]  1.005495e-03  3.040866e-04 -1.592264e-04 -6.224567e-05  4.406548e-04
#>  [7,] -1.428923e-04  1.050385e-04  2.129342e-04  1.949475e-04  6.917610e-05
#>  [8,] -2.048594e-05  2.388994e-04  4.102929e-04  3.881077e-04  1.754000e-04
#>  [9,]  2.242225e-04  2.099499e-04  1.577846e-04  1.801289e-04  2.207554e-04
#> [10,] -5.420691e-04 -1.011705e-04  1.052865e-04  1.983753e-05 -1.666568e-04
#>                 V28           V29            V3           V30           V31
#>  [1,]  5.131655e-05 -8.640852e-06 -1.681742e-05 -5.573018e-05  5.147655e-05
#>  [2,]  1.156680e-03  1.161579e-03  1.570514e-04  9.292392e-04  2.451828e-04
#>  [3,] -3.708534e-05 -8.592791e-05 -2.802225e-05 -1.238899e-04  4.861780e-06
#>  [4,]  2.056536e-04  1.913587e-04  2.398584e-05  1.354709e-04  2.710144e-05
#>  [5,]  6.354992e-04  6.383604e-04  1.125940e-04  4.197480e-04 -1.883428e-04
#>  [6,]  9.820458e-04  9.823493e-04  1.577414e-04  6.809831e-04 -1.282162e-04
#>  [7,] -1.593619e-04 -2.051996e-04 -5.035155e-05 -1.668259e-04  1.132754e-04
#>  [8,] -1.715497e-05 -9.993921e-06 -4.234225e-05  5.494033e-05  2.868870e-04
#>  [9,]  1.973932e-04  1.628885e-04  1.482211e-05  1.144482e-04  8.812406e-05
#> [10,] -5.622657e-04 -6.208560e-04 -1.268759e-04 -5.828735e-04 -4.052427e-05
#>                 V32           V33           V34           V35           V36
#>  [1,]  6.928055e-06 -2.338158e-05  9.584788e-06  5.747922e-05  1.580363e-04
#>  [2,]  3.069559e-04  2.825758e-04 -6.465408e-05 -4.628582e-04 -8.448350e-04
#>  [3,] -3.633566e-05 -6.621623e-05 -3.368147e-05  2.168360e-05  1.397723e-04
#>  [4,]  4.031648e-05  1.860845e-05 -5.475475e-05 -1.292731e-04 -1.802863e-04
#>  [5,] -1.000919e-04 -1.753669e-04 -5.825054e-04 -9.626075e-04 -1.247576e-03
#>  [6,] -1.089022e-05 -1.145148e-04 -6.634072e-04 -1.193605e-03 -1.593437e-03
#>  [7,]  5.854087e-05  5.108708e-05  2.189975e-04  3.973047e-04  5.641451e-04
#>  [8,]  2.254429e-04  2.845977e-04  4.890022e-04  6.382052e-04  7.420584e-04
#>  [9,]  8.175770e-05  5.317086e-05  2.526847e-05  1.479648e-07 -7.684265e-06
#> [10,] -1.521502e-04 -1.919509e-04  6.784233e-05  3.749289e-04  7.544751e-04
#>                 V37           V38           V39            V4           V40
#>  [1,]  1.664400e-04 -6.288260e-06 -1.050550e-04 -2.451862e-05 -1.073976e-05
#>  [2,] -7.492093e-04  2.275710e-04  7.626834e-04  2.420119e-04  3.239684e-04
#>  [3,]  1.515108e-04 -4.712769e-05 -1.484579e-04 -3.796906e-05 -3.501494e-05
#>  [4,] -1.445600e-04  2.002926e-05  1.173702e-04  3.927250e-05  5.220143e-05
#>  [5,] -1.033252e-03 -7.797599e-05  5.006754e-04  1.901694e-04  1.362218e-04
#>  [6,] -1.329067e-03 -5.727811e-05  7.045884e-04  2.605487e-04  2.124857e-04
#>  [7,]  4.904753e-04  4.376261e-05 -2.267103e-04 -8.210068e-05 -3.998520e-05
#>  [8,]  5.903540e-04  1.626903e-04 -8.413443e-05 -7.076566e-05  6.178832e-05
#>  [9,]  5.509091e-06  4.586880e-05  6.204036e-05  2.037971e-05  4.760072e-05
#> [10,]  6.849405e-04 -1.352029e-04 -5.662045e-04 -1.839093e-04 -1.790430e-04
#>                 V41           V42           V43           V44           V45
#>  [1,] -2.239877e-05 -3.218010e-05 -2.692960e-05 -8.665701e-05 -1.672341e-04
#>  [2,]  4.072008e-04  7.765786e-04  8.762091e-04  9.187173e-04  1.195290e-03
#>  [3,] -4.868965e-05 -7.761436e-05 -8.079883e-05 -1.308764e-04 -2.198112e-04
#>  [4,]  6.340412e-05  1.348153e-04  1.540387e-04  1.674973e-04  2.066079e-04
#>  [5,]  1.971675e-04  5.543590e-04  6.598178e-04  7.543250e-04  1.013983e-03
#>  [6,]  3.321261e-04  8.182141e-04  9.446309e-04  1.064322e-03  1.415719e-03
#>  [7,] -7.855268e-05 -2.074954e-04 -2.410461e-04 -3.068293e-04 -4.581919e-04
#>  [8,]  2.173155e-05 -1.563265e-04 -2.100579e-04 -2.658270e-04 -3.981445e-04
#>  [9,]  5.617066e-05  1.002645e-04  1.086629e-04  9.552439e-05  9.657377e-05
#> [10,] -2.607999e-04 -5.118095e-04 -5.587017e-04 -6.635538e-04 -9.660742e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -1.518267e-04 -1.212435e-04 -8.379892e-05 -4.267696e-05  8.322772e-06
#>  [2,]  9.915966e-04  7.161303e-04  5.331006e-04  3.260505e-04  2.004760e-04
#>  [3,] -1.924905e-04 -1.490355e-04 -1.105819e-04 -5.827274e-05 -4.091124e-06
#>  [4,]  1.683407e-04  1.221880e-04  8.864648e-05  5.523509e-05  3.808900e-05
#>  [5,]  8.426676e-04  6.203261e-04  4.403928e-04  2.780975e-04  1.591756e-04
#>  [6,]  1.175979e-03  8.590396e-04  6.108535e-04  3.819320e-04  2.195310e-04
#>  [7,] -3.879975e-04 -2.906518e-04 -2.063085e-04 -1.239626e-04 -4.610038e-05
#>  [8,] -3.432276e-04 -2.544712e-04 -1.761764e-04 -1.128043e-04 -5.448175e-05
#>  [9,]  7.509076e-05  4.820628e-05  3.787993e-05  2.474982e-05  3.150092e-05
#> [10,] -8.205315e-04 -6.098682e-04 -4.574093e-04 -2.626308e-04 -1.061618e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -1.183232e-05 -1.548125e-05 -4.224806e-06 -1.940025e-06 -4.979392e-06
#>  [2,]  8.150326e-05  1.087814e-04  6.137924e-05  2.696846e-05  4.380034e-05
#>  [3,] -1.550203e-05 -2.203688e-05 -8.879976e-06 -4.369367e-06 -7.705232e-06
#>  [4,]  1.347681e-05  1.749138e-05  9.546534e-06  3.867701e-06  6.883735e-06
#>  [5,]  6.386730e-05  8.478806e-05  4.353212e-05  1.441382e-05  3.061537e-05
#>  [6,]  8.878830e-05  1.193541e-04  6.320285e-05  2.218637e-05  4.313943e-05
#>  [7,] -2.948478e-05 -3.930228e-05 -1.826844e-05 -6.545247e-06 -1.336610e-05
#>  [8,] -2.201906e-05 -3.362076e-05 -1.774617e-05 -2.251147e-06 -9.616391e-06
#>  [9,]  6.540714e-06  8.553463e-06  7.143219e-06  2.997938e-06  4.152775e-06
#> [10,] -6.644842e-05 -9.059847e-05 -4.702442e-05 -1.950779e-05 -3.320596e-05
#>                 V55           V56           V57           V58           V59
#>  [1,] -8.158981e-06 -4.672101e-06 -7.657082e-07 -1.038283e-05 -4.370348e-06
#>  [2,]  3.739137e-05  2.505713e-05  1.311570e-05  5.081049e-05  3.065130e-05
#>  [3,] -9.953815e-06 -6.497068e-06 -2.295249e-06 -1.251289e-05 -6.270256e-06
#>  [4,]  5.902644e-06  3.772572e-06  1.496210e-06  8.164045e-06  4.464610e-06
#>  [5,]  2.716855e-05  1.569262e-05  2.778776e-06  4.135367e-05  2.016739e-05
#>  [6,]  3.810887e-05  2.243570e-05  5.525195e-06  5.596364e-05  2.882600e-05
#>  [7,] -1.415039e-05 -8.138923e-06 -1.425322e-06 -2.061214e-05 -9.809658e-06
#>  [8,] -7.307275e-06 -3.052198e-06  2.022410e-06 -1.371995e-05 -5.842101e-06
#>  [9,]  1.925849e-06  1.614776e-06  1.846810e-06  2.076767e-06  2.464331e-06
#> [10,] -3.321281e-05 -2.182522e-05 -9.043638e-06 -4.422287e-05 -2.448018e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -2.011437e-05 -1.648008e-06 -7.297222e-05 -1.027277e-04 -1.541181e-04
#>  [2,]  3.048111e-04  1.531959e-05  4.584004e-04  5.783050e-04  9.608090e-04
#>  [3,] -3.695313e-05 -2.826640e-06 -9.474720e-05 -1.336993e-04 -2.016323e-04
#>  [4,]  5.357991e-05  2.142312e-06  7.769612e-05  9.700503e-05  1.639719e-04
#>  [5,]  2.372930e-04  8.124427e-06  3.697113e-04  4.680703e-04  8.080607e-04
#>  [6,]  3.218775e-04  1.175135e-05  4.940981e-04  6.290489e-04  1.093045e-03
#>  [7,] -8.637166e-05 -3.744525e-06 -1.632204e-04 -2.150050e-04 -3.671080e-04
#>  [8,] -7.470997e-05 -1.150116e-06 -1.184627e-04 -1.610945e-04 -3.067215e-04
#>  [9,]  3.297238e-05  1.526312e-06  2.974530e-05  3.465422e-05  6.703844e-05
#> [10,] -2.000762e-04 -1.172116e-05 -3.681660e-04 -4.918339e-04 -8.122657e-04
#> 
#> $vW[[2]]
#>             [,1]          [,2]        [,3]         [,4]         [,5]
#> [1,] -0.01001619 -0.0001736436 -0.01166444  0.006645511  0.007236953
#> [2,]  0.01001619  0.0001736436  0.01166444 -0.006645511 -0.007236953
#>              [,6]          [,7]        [,8]         [,9]       [,10]
#> [1,]  0.007839383  0.0009708955  0.01011112  0.002472973 -0.00787849
#> [2,] -0.007839383 -0.0009708955 -0.01011112 -0.002472973  0.00787849
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.6367958  0.5987279  0.8151084  0.2738067  0.3221524  0.1579607
#>  [7]  0.4493566 -0.1360731  0.4156402  0.6934822
#> 
#> $B[[2]]
#> [1] -0.0003912439 -0.1045143901
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  1.600988e-05  1.215375e-03 -1.053141e-04  1.911775e-04  3.957703e-04
#>  [6]  6.910041e-04 -9.445242e-05  2.511542e-04  2.116232e-04 -6.768232e-04
#> 
#> $vB[[2]]
#> [1] -0.01996017  0.01996017
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#>  [2,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#>  [3,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#>  [4,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#>  [5,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#>  [6,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#>  [7,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [8,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#>  [9,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [10,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [11,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [12,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [13,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [14,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [15,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [16,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [17,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [18,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [19,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [20,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [21,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [22,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [23,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [24,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [25,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [26,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [27,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [28,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [29,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [30,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [31,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [32,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [33,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [34,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [35,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [36,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [37,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [38,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [39,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [40,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#>  [2,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#>  [3,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#>  [4,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#>  [5,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#>  [6,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#>  [7,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [8,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#>  [9,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [10,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [11,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [12,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [13,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [14,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [15,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [16,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [17,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [18,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [19,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [20,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [21,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [22,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [23,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [24,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [25,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [26,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [27,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [28,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [29,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [30,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [31,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [32,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [33,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [34,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [35,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [36,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [37,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [38,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [39,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [40,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#>  [2,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#>  [3,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#>  [4,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#>  [5,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#>  [6,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#>  [7,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [8,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#>  [9,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [10,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [11,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [12,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [13,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [14,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [15,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [16,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [17,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [18,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [19,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [20,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [21,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [22,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [23,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [24,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [25,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [26,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [27,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [28,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [29,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [30,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [31,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [32,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [33,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [34,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [35,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [36,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [37,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [38,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [39,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [40,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#>  [2,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#>  [3,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#>  [4,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#>  [5,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#>  [6,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#>  [7,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [8,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#>  [9,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [10,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [11,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [12,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [13,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [14,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [15,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [16,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [17,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [18,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [19,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [20,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [21,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [22,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [23,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [24,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [25,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [26,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [27,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [28,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [29,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [30,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [31,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [32,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [33,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [34,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [35,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [36,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [37,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [38,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [39,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [40,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#>  [2,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#>  [3,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#>  [4,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#>  [5,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#>  [6,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#>  [7,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [8,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#>  [9,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [10,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [11,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [12,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [13,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [14,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [15,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [16,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [17,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [18,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [19,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [20,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [21,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [22,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [23,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [24,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [25,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [26,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [27,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [28,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [29,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [30,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [31,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [32,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [33,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [34,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [35,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [36,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [37,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [38,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [39,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [40,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#>  [2,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#>  [3,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#>  [4,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#>  [5,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#>  [6,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#>  [7,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [8,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#>  [9,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [10,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [11,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [12,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [13,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [14,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [15,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [16,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [17,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [18,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [19,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [20,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [21,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [22,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [23,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [24,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [25,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [26,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [27,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [28,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [29,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [30,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [31,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [32,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [33,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [34,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [35,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [36,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [37,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [38,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [39,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [40,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8163985 0.6157091 0.8573164 0.5036654 0.4200234 0.4563517 0.6284492
#>  [2,] 0.8560678 0.7478056 0.9151245 0.5663539 0.5056911 0.5978734 0.7145939
#>  [3,] 0.7535200 0.6221028 0.8130480 0.3923266 0.3835998 0.4193682 0.5290710
#>  [4,] 0.6970649 0.6145280 0.7645745 0.4315609 0.3752102 0.3809497 0.5238721
#>  [5,] 0.7736843 0.7487800 0.8587912 0.5043406 0.4872538 0.5354796 0.6125334
#>  [6,] 0.5967964 0.5487543 0.7113081 0.2836505 0.2706018 0.2720909 0.3751612
#>  [7,] 0.7907734 0.6734161 0.8569362 0.4588259 0.4020581 0.4905462 0.6218168
#>  [8,] 0.8291002 0.7379171 0.8948774 0.5612615 0.5093428 0.5803651 0.7103964
#>  [9,] 0.8182518 0.7100891 0.8818579 0.4854665 0.4377015 0.5149140 0.6056471
#> [10,] 0.7334597 0.7052719 0.8446399 0.3489456 0.3925836 0.4094705 0.5260013
#> [11,] 0.7693258 0.7509694 0.8475844 0.5488912 0.5434707 0.5894150 0.6587437
#> [12,] 0.8423988 0.7337678 0.9051930 0.5410668 0.5124278 0.5718690 0.6862058
#> [13,] 0.8345666 0.7562781 0.9029052 0.5658662 0.5251743 0.6080843 0.7031197
#> [14,] 0.5988448 0.5725577 0.7031285 0.2650233 0.2577426 0.2642253 0.3583963
#> [15,] 0.7776244 0.6615470 0.8472143 0.4420364 0.3512406 0.4313729 0.5492250
#> [16,] 0.7777336 0.7749604 0.8756270 0.5051406 0.4898088 0.5808774 0.6469200
#> [17,] 0.8021199 0.7933093 0.8945810 0.5074875 0.4857721 0.5707919 0.6397610
#> [18,] 0.7650027 0.7522195 0.8452241 0.5386434 0.5268797 0.5736238 0.6467197
#> [19,] 0.8376942 0.7525085 0.9050490 0.5541401 0.5212418 0.5978609 0.6995367
#> [20,] 0.7690090 0.6175674 0.8141969 0.4936612 0.4580265 0.4367048 0.5938428
#> [21,] 0.8268921 0.6813868 0.8713405 0.5883990 0.5604619 0.5711129 0.7044518
#> [22,] 0.7982794 0.6710915 0.8624280 0.5123214 0.4885746 0.5104900 0.6405891
#> [23,] 0.7950576 0.6987835 0.8631688 0.5434130 0.5067078 0.5410304 0.6772844
#> [24,] 0.8625397 0.7625110 0.9210931 0.6162271 0.5631919 0.6377641 0.7404921
#> [25,] 0.7049316 0.7336248 0.8165381 0.4767702 0.4569574 0.4953963 0.5792203
#> [26,] 0.7952242 0.7438265 0.8705491 0.4811695 0.4894771 0.5319629 0.6222015
#> [27,] 0.6172230 0.7010140 0.7396820 0.3273794 0.3333340 0.3652274 0.4295896
#> [28,] 0.8685623 0.7668118 0.9248657 0.6268652 0.5622581 0.6476053 0.7496499
#> [29,] 0.7773752 0.7873175 0.8783998 0.5034084 0.4743552 0.5861650 0.6538808
#> [30,] 0.7466362 0.6351319 0.8094185 0.4085907 0.4094962 0.4262668 0.5271000
#> [31,] 0.6840433 0.7075045 0.8229689 0.3189697 0.3224856 0.4074084 0.4866259
#> [32,] 0.5687106 0.6073493 0.7010992 0.2497462 0.2497410 0.3113095 0.3799536
#> [33,] 0.8047505 0.6967611 0.8811978 0.4650677 0.4366365 0.5006265 0.5986183
#> [34,] 0.8190393 0.7731942 0.8981238 0.5602652 0.5285317 0.6264070 0.6868443
#> [35,] 0.7472490 0.7639682 0.8578216 0.4352458 0.4278648 0.4911866 0.5802858
#> [36,] 0.8407087 0.8200719 0.9188219 0.5642615 0.5406232 0.6329875 0.7031126
#> [37,] 0.8098482 0.8072638 0.8956896 0.5337132 0.5084752 0.6110360 0.6767987
#> [38,] 0.7977909 0.6580983 0.8528437 0.4808256 0.4449608 0.4536477 0.6089478
#> [39,] 0.7598023 0.7640186 0.8496616 0.4760864 0.4649396 0.5235289 0.5990520
#> [40,] 0.8562081 0.7459939 0.9122138 0.6137069 0.5686116 0.6283683 0.7378246
#>            [,8]      [,9]     [,10]
#>  [1,] 0.1671008 0.6429367 0.7952193
#>  [2,] 0.2049881 0.7406559 0.8611943
#>  [3,] 0.1841822 0.6479454 0.7202815
#>  [4,] 0.1969170 0.5416373 0.6919685
#>  [5,] 0.2774977 0.7352635 0.8051438
#>  [6,] 0.1483986 0.5249385 0.5921045
#>  [7,] 0.1877689 0.6931418 0.7897772
#>  [8,] 0.2231362 0.7057536 0.8414794
#>  [9,] 0.2091663 0.6881622 0.8208168
#> [10,] 0.1660076 0.6853242 0.7897025
#> [11,] 0.3574905 0.7296890 0.8043368
#> [12,] 0.2131131 0.7405030 0.8508313
#> [13,] 0.2573402 0.7238326 0.8408751
#> [14,] 0.1541663 0.5325054 0.6070939
#> [15,] 0.1532322 0.6679951 0.7662462
#> [16,] 0.2944433 0.7224687 0.8162099
#> [17,] 0.2579363 0.7594123 0.8327799
#> [18,] 0.3413824 0.7192356 0.7996622
#> [19,] 0.2375526 0.7276374 0.8404602
#> [20,] 0.1959223 0.6244398 0.7597322
#> [21,] 0.2518243 0.7057808 0.8199956
#> [22,] 0.2083383 0.6862522 0.7942553
#> [23,] 0.2214162 0.6810931 0.8103133
#> [24,] 0.2276598 0.7662362 0.8751799
#> [25,] 0.2957428 0.6776635 0.7316600
#> [26,] 0.2447477 0.7356504 0.8324941
#> [27,] 0.2409185 0.6406639 0.6888409
#> [28,] 0.2289091 0.7672483 0.8789721
#> [29,] 0.2895343 0.7305762 0.8254281
#> [30,] 0.1950364 0.6620550 0.7709763
#> [31,] 0.1915386 0.6568723 0.7595185
#> [32,] 0.1741038 0.5385017 0.6664067
#> [33,] 0.2046536 0.7022919 0.8099371
#> [34,] 0.2834994 0.7364420 0.8287419
#> [35,] 0.2304362 0.6939299 0.8007512
#> [36,] 0.2750146 0.7922792 0.8617457
#> [37,] 0.2933424 0.7745308 0.8437161
#> [38,] 0.1886253 0.6823723 0.8009828
#> [39,] 0.2793393 0.7356961 0.8006330
#> [40,] 0.2395931 0.7553313 0.8658205
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5299756 0.4700244
#>  [2,] 0.5249883 0.4750117
#>  [3,] 0.5276217 0.4723783
#>  [4,] 0.5269268 0.4730732
#>  [5,] 0.5237617 0.4762383
#>  [6,] 0.5275773 0.4724227
#>  [7,] 0.5278516 0.4721484
#>  [8,] 0.5246843 0.4753157
#>  [9,] 0.5269302 0.4730698
#> [10,] 0.5280000 0.4720000
#> [11,] 0.5219017 0.4780983
#> [12,] 0.5252221 0.4747779
#> [13,] 0.5234773 0.4765227
#> [14,] 0.5278453 0.4721547
#> [15,] 0.5288442 0.4711558
#> [16,] 0.5229286 0.4770714
#> [17,] 0.5233632 0.4766368
#> [18,] 0.5222042 0.4777958
#> [19,] 0.5237877 0.4762123
#> [20,] 0.5274153 0.4725847
#> [21,] 0.5243294 0.4756706
#> [22,] 0.5258573 0.4741427
#> [23,] 0.5247878 0.4752122
#> [24,] 0.5230665 0.4769335
#> [25,] 0.5223525 0.4776475
#> [26,] 0.5250196 0.4749804
#> [27,] 0.5247888 0.4752112
#> [28,] 0.5229985 0.4770015
#> [29,] 0.5231779 0.4768221
#> [30,] 0.5280259 0.4719741
#> [31,] 0.5278164 0.4721836
#> [32,] 0.5278788 0.4721212
#> [33,] 0.5271751 0.4728249
#> [34,] 0.5219594 0.4780406
#> [35,] 0.5249950 0.4750050
#> [36,] 0.5219424 0.4780576
#> [37,] 0.5224898 0.4775102
#> [38,] 0.5284784 0.4715216
#> [39,] 0.5238233 0.4761767
#> [40,] 0.5232744 0.4767256
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]      [,3]        [,4]        [,5]         [,6]
#>  [1,] 1.4921352 0.4713747 1.7931773  0.01466206 -0.32267739 -0.175038605
#>  [2,] 1.7830074 1.0869430 2.3778752  0.26699029  0.02276545  0.396612229
#>  [3,] 1.1174746 0.4984831 1.4699382 -0.43754282 -0.47429662 -0.325367463
#>  [4,] 0.8333600 0.4663860 1.1779250 -0.27548571 -0.50992883 -0.485519367
#>  [5,] 1.2292330 1.0921163 1.8052859  0.01736271 -0.05099599  0.142157516
#>  [6,] 0.3921343 0.1956390 0.9017458 -0.92642550 -0.99157136 -0.984039751
#>  [7,] 1.3295935 0.7236767 1.7900730 -0.16507036 -0.39689703 -0.037819899
#>  [8,] 1.5792634 1.0351706 2.1415595  0.24628330  0.03737554  0.324272377
#>  [9,] 1.5045477 0.8958168 2.0101427 -0.05815035 -0.25049584  0.059673559
#> [10,] 1.0122473 0.8725300 1.6931649 -0.62367711 -0.43646523 -0.366154845
#> [11,] 1.2045080 1.1037890 1.7157795  0.19619151  0.17432289  0.361547606
#> [12,] 1.6761861 1.0138236 2.2563050  0.16463791  0.04972133  0.289480596
#> [13,] 1.6183439 1.1323811 2.2299297  0.26500482  0.10078253  0.439266871
#> [14,] 0.4006539 0.2922940 0.8622405 -1.02002095 -1.05773487 -1.024121723
#> [15,] 1.2518759 0.6701960 1.7129170 -0.23290148 -0.61359060 -0.276252008
#> [16,] 1.2525074 1.2365357 1.9516552  0.02056307 -0.04077059  0.326376312
#> [17,] 1.3995966 1.3449897 2.1384122  0.02995235 -0.05692695  0.285082940
#> [18,] 1.1803051 1.1104851 1.6976239  0.15488265  0.10762260  0.296651699
#> [19,] 1.6411712 1.1120360 2.2546278  0.21741259  0.08501846  0.396560215
#> [20,] 1.2027241 0.4792359 1.4775145 -0.02535656 -0.16829009 -0.254546177
#> [21,] 1.5637589 0.7601525 1.9128632  0.35735082  0.24303673  0.286393366
#> [22,] 1.3755748 0.7131259 1.8356040  0.04929544 -0.04570948  0.041966149
#> [23,] 1.3556854 0.8415116 1.8418619  0.17409053  0.02683295  0.164491351
#> [24,] 1.8365464 1.1664958 2.4572929  0.47356475  0.25412639  0.565672892
#> [25,] 0.8708934 1.0130919 1.4930666 -0.09298618 -0.17259770 -0.018415411
#> [26,] 1.3567085 1.0659530 1.9058221 -0.07535763 -0.04209780  0.128026052
#> [27,] 0.4777778 0.8521309 1.0443166 -0.72006158 -0.69314439 -0.552746768
#> [28,] 1.8883064 1.1903956 2.5103712  0.51879182  0.25033171  0.608529342
#> [29,] 1.2504354 1.3088311 1.9773629  0.01363398 -0.10266927  0.348133974
#> [30,] 1.0807515 0.5542969 1.4462359 -0.36979444 -0.36604851 -0.297098843
#> [31,] 0.7724158 0.8832946 1.5365933 -0.75851059 -0.74237251 -0.374689374
#> [32,] 0.2765922 0.4361835 0.8525376 -1.09996628 -1.09999412 -0.794004215
#> [33,] 1.4162540 0.8319215 2.0038219 -0.13995702 -0.25482392  0.002505908
#> [34,] 1.5098523 1.2264359 2.1765495  0.24223859  0.11425088  0.516833389
#> [35,] 1.0839937 1.1745595 1.7973134 -0.26047982 -0.29056794 -0.035257433
#> [36,] 1.6635108 1.5168347 2.4264474  0.25847534  0.16285194  0.545054775
#> [37,] 1.4490241 1.4323283 2.1502233  0.13505755  0.03390412  0.451669153
#> [38,] 1.3725445 0.6548309 1.7570813 -0.07673514 -0.22105268 -0.185942965
#> [39,] 1.1515958 1.1748389 1.7319492 -0.09572746 -0.14047232  0.094185300
#> [40,] 1.7841465 1.0773592 2.3409702  0.46292122  0.27618859  0.525223202
#>              [,7]       [,8]       [,9]     [,10]
#>  [1,]  0.52556978 -1.6063153 0.58813339 1.3566784
#>  [2,]  0.91780152 -1.3554049 1.04938064 1.8252451
#>  [3,]  0.11641519 -1.4882655 0.61002002 0.9458583
#>  [4,]  0.09556105 -1.4056755 0.16693571 0.8093382
#>  [5,]  0.45797367 -0.9569081 1.02149385 1.4187590
#>  [6,] -0.51013773 -1.7472169 0.09983692 0.3726723
#>  [7,]  0.49726695 -1.4645732 0.81484905 1.3235828
#>  [8,]  0.89731005 -1.2474828 0.87484859 1.6692771
#>  [9,]  0.42905114 -1.3299583 0.79154143 1.5218909
#> [10,]  0.10409906 -1.6141910 0.77834907 1.3231334
#> [11,]  0.65770070 -0.5862730 0.99304542 1.4136230
#> [12,]  0.78244037 -1.3062613 1.04858470 1.7411359
#> [13,]  0.86219826 -1.0598393 0.96355283 1.6647539
#> [14,] -0.58233133 -1.7022906 0.13020510 0.4351130
#> [15,]  0.19753994 -1.7094721 0.69913096 1.1872349
#> [16,]  0.60552800 -0.8739008 0.95674027 1.4908772
#> [17,]  0.57432712 -1.0567224 1.14946007 1.6054585
#> [18,]  0.60465133 -0.6571398 0.94067330 1.3841847
#> [19,]  0.84509245 -1.1661445 0.98266877 1.6616564
#> [20,]  0.37987438 -1.4119777 0.50843615 1.1512121
#> [21,]  0.86858791 -1.0889063 0.87497977 1.5163175
#> [22,]  0.57792187 -1.3349711 0.78265587 1.3507687
#> [23,]  0.74132002 -1.2574327 0.75879972 1.4520471
#> [24,]  1.04852781 -1.2215726 1.18717910 1.9475557
#> [25,]  0.31957342 -0.8676533 0.74305490 1.0030611
#> [26,]  0.49890304 -1.1268241 1.02348261 1.6034077
#> [27,] -0.28352584 -1.1476504 0.57824666 0.7947060
#> [28,]  1.09674587 -1.2144816 1.19283840 1.9827324
#> [29,]  0.63614187 -0.8976470 0.99754783 1.5535656
#> [30,]  0.10850644 -1.4176109 0.67246547 1.2138324
#> [31,] -0.05350910 -1.4400439 0.64938704 1.1500417
#> [32,] -0.48974514 -1.5568173 0.15431240 0.6919775
#> [33,]  0.39971152 -1.3574590 0.85823580 1.4496015
#> [34,]  0.78540723 -0.9271690 1.02755715 1.5767371
#> [35,]  0.32394684 -1.2058497 0.81855666 1.3909960
#> [36,]  0.86216428 -0.9693275 1.33871890 1.8298654
#> [37,]  0.73909836 -0.8792059 1.23407394 1.6861416
#> [38,]  0.44289137 -1.4589677 0.76469567 1.3924485
#> [39,]  0.40151646 -0.9477412 1.02371742 1.3902550
#> [40,]  1.03469251 -1.1549115 1.12725125 1.8644988
#> 
#> $pre[[3]]
#>              [,1]       [,2]
#>  [1,] -0.07077812 -0.1908244
#>  [2,] -0.08370858 -0.1837451
#>  [3,] -0.06177930 -0.1723786
#>  [4,] -0.06257512 -0.1703867
#>  [5,] -0.07497025 -0.1700887
#>  [6,] -0.04883650 -0.1592578
#>  [7,] -0.07015688 -0.1816787
#>  [8,] -0.08180839 -0.1806258
#>  [9,] -0.07260192 -0.1804269
#> [10,] -0.06043715 -0.1725543
#> [11,] -0.07773412 -0.1653970
#> [12,] -0.08030291 -0.1812769
#> [13,] -0.08255572 -0.1765340
#> [14,] -0.04714000 -0.1586364
#> [15,] -0.06750558 -0.1830106
#> [16,] -0.07633308 -0.1681118
#> [17,] -0.07803276 -0.1715537
#> [18,] -0.07688357 -0.1657589
#> [19,] -0.08195377 -0.1771765
#> [20,] -0.06949961 -0.1792707
#> [21,] -0.08178831 -0.1791826
#> [22,] -0.07480373 -0.1783253
#> [23,] -0.07859292 -0.1778255
#> [24,] -0.08931352 -0.1816450
#> [25,] -0.07114849 -0.1606180
#> [26,] -0.07351524 -0.1736774
#> [27,] -0.05615804 -0.1553947
#> [28,] -0.09048022 -0.1825391
#> [29,] -0.07677695 -0.1695550
#> [30,] -0.06301432 -0.1752357
#> [31,] -0.05716345 -0.1685440
#> [32,] -0.04742137 -0.1590522
#> [33,] -0.07074142 -0.1795489
#> [34,] -0.08280584 -0.1706998
#> [35,] -0.06942781 -0.1694912
#> [36,] -0.08491584 -0.1727418
#> [37,] -0.08067044 -0.1706902
#> [38,] -0.06991543 -0.1839526
#> [39,] -0.07261847 -0.1679840
#> [40,] -0.08790131 -0.1810664
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4700244 -0.4700244
#>  [2,]  0.4750117 -0.4750117
#>  [3,]  0.4723783 -0.4723783
#>  [4,]  0.4730732 -0.4730732
#>  [5,] -0.5237617  0.5237617
#>  [6,]  0.4724227 -0.4724227
#>  [7,] -0.5278516  0.5278516
#>  [8,]  0.4753157 -0.4753157
#>  [9,]  0.4730698 -0.4730698
#> [10,]  0.4720000 -0.4720000
#> [11,] -0.5219017  0.5219017
#> [12,]  0.4747779 -0.4747779
#> [13,]  0.4765227 -0.4765227
#> [14,]  0.4721547 -0.4721547
#> [15,] -0.5288442  0.5288442
#> [16,]  0.4770714 -0.4770714
#> [17,] -0.5233632  0.5233632
#> [18,] -0.5222042  0.5222042
#> [19,]  0.4762123 -0.4762123
#> [20,]  0.4725847 -0.4725847
#> [21,] -0.5243294  0.5243294
#> [22,] -0.5258573  0.5258573
#> [23,]  0.4752122 -0.4752122
#> [24,] -0.5230665  0.5230665
#> [25,] -0.5223525  0.5223525
#> [26,] -0.5250196  0.5250196
#> [27,] -0.5247888  0.5247888
#> [28,] -0.5229985  0.5229985
#> [29,] -0.5231779  0.5231779
#> [30,] -0.5280259  0.5280259
#> [31,]  0.4721836 -0.4721836
#> [32,]  0.4721212 -0.4721212
#> [33,]  0.4728249 -0.4728249
#> [34,]  0.4780406 -0.4780406
#> [35,]  0.4750050 -0.4750050
#> [36,] -0.5219424  0.5219424
#> [37,] -0.5224898  0.5224898
#> [38,]  0.4715216 -0.4715216
#> [39,] -0.5238233  0.5238233
#> [40,] -0.5232744  0.5232744
#> 
#> $L
#> [1] 0.6961662 0.6963223 0.6938587 0.6927065 0.6983709 0.6898560
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4057971