Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12          V13          V14
#>  [1,] -0.4099287 -0.09643719 -0.064657654 -0.01902087 -0.138217852 -0.162163230
#>  [2,] -0.2910013 -0.11740335 -0.150330457 -0.15233286 -0.067204649 -0.107175083
#>  [3,] -0.2242811 -0.10291955 -0.198451971 -0.23853106 -0.227496366 -0.184771063
#>  [4,] -0.2423015 -0.18237264 -0.168901314 -0.04913330 -0.066120479 -0.052900862
#>  [5,] -0.2359666 -0.23407889 -0.099730775 -0.16351872 -0.186020436 -0.051264753
#>  [6,] -0.2559627 -0.04847141 -0.002151581 -0.02191670  0.009576282 -0.072678057
#>  [7,] -0.2411615 -0.11828171 -0.128011407 -0.02389749 -0.141195845 -0.009604338
#>  [8,] -0.3204049 -0.15904194 -0.155493805 -0.15749398 -0.048974051 -0.037835731
#>  [9,] -0.3128532 -0.19053386 -0.178469479 -0.10372313 -0.031936038 -0.107625790
#> [10,] -0.4056913 -0.12591076 -0.129199905 -0.10591183 -0.088939250  0.018449179
#>               V15         V16          V17          V18         V19         V2
#>  [1,] -0.02945589 -0.01638806 -0.049959516 -0.093447728  0.07529761 -0.2798157
#>  [2,] -0.02153476 -0.02292220 -0.066737794 -0.107889170 -0.04080386 -0.3266469
#>  [3,] -0.04365652 -0.11379626 -0.097451729  0.003775424  0.07962267 -0.2215124
#>  [4,] -0.02667964 -0.03816047  0.109996614 -0.048439876  0.06293548 -0.2137509
#>  [5,] -0.16242686 -0.11742794 -0.049755066 -0.077734012 -0.05816081 -0.3426149
#>  [6,] -0.02483156 -0.08559417  0.023522628  0.051635459  0.10063109 -0.3134112
#>  [7,] -0.07651928  0.06353208  0.046844603  0.013547597 -0.01293253 -0.3480731
#>  [8,] -0.02880281 -0.07734433 -0.023689816  0.033260633 -0.06451062 -0.3198878
#>  [9,] -0.16838320  0.03894360  0.038580774  0.065380866 -0.03052405 -0.3755247
#> [10,] -0.03208247 -0.02732946  0.003265397  0.142841317  0.03811481 -0.2395060
#>                V20          V21         V22          V23          V24
#>  [1,]  0.104233012  0.116685671  0.04848944  0.080535825  0.126281664
#>  [2,] -0.056182135 -0.023530142  0.10722888  0.217810337  0.262305398
#>  [3,] -0.042089861 -0.002328364  0.09061335  0.061440469  0.133807910
#>  [4,] -0.015931527  0.130327386 -0.02323388 -0.009720257 -0.003917172
#>  [5,] -0.008785723 -0.045556071 -0.01184390 -0.037036208  0.009866882
#>  [6,]  0.161091886  0.210443119  0.11221837  0.159109859  0.222814672
#>  [7,] -0.004845350  0.118192578  0.18810345  0.242505139  0.113388494
#>  [8,]  0.013430335 -0.092648349  0.03369805 -0.009360310  0.028788713
#>  [9,]  0.097728387  0.007352089  0.01725495  0.152090826  0.113190744
#> [10,]  0.106998995  0.184060807  0.11966678  0.089115654  0.272578804
#>              V25         V26         V27         V28         V29         V3
#>  [1,] 0.06916859 0.131301154  0.05084945  0.07552822  0.02656281 -0.2288506
#>  [2,] 0.19674509 0.136460248  0.21435348  0.07545861  0.15090409 -0.2963678
#>  [3,] 0.08320916 0.003187627  0.17491114  0.03996741  0.03607942 -0.2804642
#>  [4,] 0.04935730 0.041599767 -0.04393143  0.08270093 -0.08057532 -0.2353605
#>  [5,] 0.14372672 0.217433027  0.20859356 -0.01931877 -0.01887486 -0.1842968
#>  [6,] 0.16910290 0.280870301  0.08918962  0.08204238  0.20415512 -0.3222988
#>  [7,] 0.13033780 0.203609085  0.27214265  0.29169635  0.19826210 -0.3818310
#>  [8,] 0.20591714 0.183696004  0.17257625  0.19418716  0.04309916 -0.2866111
#>  [9,] 0.17059836 0.089145667  0.24827834  0.21652773  0.12584167 -0.2730438
#> [10,] 0.13684648 0.139072819  0.06585318  0.22723558  0.21793908 -0.2340874
#>                V30         V31         V32          V33         V34         V35
#>  [1,]  0.163814742  0.09159962  0.08395959 -0.011930223 -0.01317087 -0.04474762
#>  [2,]  0.183320902  0.03118879  0.11382584  0.060963366  0.06612771  0.11205503
#>  [3,]  0.189301580  0.02167686 -0.07204058  0.023609773 -0.06198573 -0.14690553
#>  [4,] -0.058828936  0.01240143 -0.15966898 -0.141679403 -0.06041914 -0.06723768
#>  [5,] -0.070033198  0.08921511  0.04465572 -0.089254683 -0.08267177 -0.08489126
#>  [6,] -0.007078984  0.01859155 -0.03851818  0.092021077 -0.03666405 -0.02434587
#>  [7,]  0.220602275  0.11266685  0.01113725 -0.022693990  0.11751871  0.07814211
#>  [8,]  0.082426287  0.10977437  0.12329259 -0.008008803  0.06944531  0.05264011
#>  [9,]  0.001844598 -0.03117185 -0.09921937 -0.024650894 -0.16351867 -0.01121946
#> [10,]  0.078662522  0.07095301  0.04052512  0.096581578  0.07129778  0.02061527
#>               V36          V37          V38          V39         V4
#>  [1,]  0.07121924  0.019761817  0.065519118  0.070119982 -0.2854502
#>  [2,]  0.01851187  0.098427592 -0.043002039 -0.105308670 -0.2695244
#>  [3,] -0.08453773 -0.053602911 -0.017776057 -0.141504505 -0.2626460
#>  [4,] -0.13622890  0.006004596  0.006244451  0.014697351 -0.2270935
#>  [5,] -0.08957759 -0.125689348 -0.203173258 -0.128858967 -0.3271626
#>  [6,] -0.05042485 -0.052218376 -0.022454088 -0.009990471 -0.3214329
#>  [7,] -0.03260901 -0.092198243 -0.003601904 -0.084698249 -0.2153525
#>  [8,]  0.05379836 -0.068081178  0.037882292 -0.075824559 -0.2089116
#>  [9,] -0.02728516 -0.131868212 -0.128179681 -0.199598617 -0.2723317
#> [10,] -0.08146496 -0.065630544 -0.127294310 -0.049723834 -0.2977114
#>                V40         V41           V42         V43          V44
#>  [1,] -0.099957948  0.04353269 -0.1512135061 -0.03504929 -0.007858075
#>  [2,]  0.044824001 -0.10169597 -0.1514163780 -0.03273046 -0.011598221
#>  [3,] -0.070809038 -0.09905274 -0.0002915394 -0.06238791 -0.134527767
#>  [4,]  0.024432278 -0.09558207 -0.0978635423 -0.14774905 -0.209117507
#>  [5,] -0.157406307 -0.20707532 -0.1516727129 -0.07685945 -0.241534366
#>  [6,] -0.143505358 -0.12061198 -0.1645980554 -0.06007376 -0.050711783
#>  [7,] -0.123806177 -0.08283647 -0.0263797036 -0.13002990 -0.060662708
#>  [8,] -0.071080597 -0.08456790 -0.0317767631 -0.19987616 -0.211156439
#>  [9,] -0.036407230 -0.00266994 -0.1998826082 -0.09687841 -0.218138730
#> [10,]  0.003980914 -0.07475393 -0.0444420191 -0.15624029 -0.059904251
#>               V45         V46         V47        V48        V49         V5
#>  [1,] -0.16502732 -0.13989237 -0.24674418 -0.2282480 -0.2573656 -0.2145391
#>  [2,] -0.07217963 -0.22052478 -0.13082728 -0.2599197 -0.3407568 -0.3061808
#>  [3,] -0.15230915 -0.24981187 -0.11337316 -0.1898996 -0.2486764 -0.2712861
#>  [4,] -0.21856422 -0.10238497 -0.19461128 -0.1412080 -0.2451768 -0.3017293
#>  [5,] -0.07786947 -0.20516791 -0.08157182 -0.2648282 -0.3182228 -0.1512535
#>  [6,] -0.18700879 -0.19411186 -0.26025885 -0.3331387 -0.2384046 -0.2064736
#>  [7,] -0.03363846 -0.09560731 -0.18870124 -0.1991384 -0.2301443 -0.2837497
#>  [8,] -0.12015565 -0.06619085 -0.11428164 -0.3089253 -0.2689693 -0.2517332
#>  [9,] -0.17308737 -0.26484376 -0.24774517 -0.1591290 -0.3081711 -0.2157345
#> [10,] -0.04457362 -0.06266637 -0.19002191 -0.2317713 -0.2854246 -0.2940952
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.4045799 -0.3479860 -0.3640204 -0.4019073 -0.4005843 -0.3416572
#>  [2,] -0.4128382 -0.3162423 -0.2579407 -0.4179856 -0.4178923 -0.3907085
#>  [3,] -0.2259053 -0.2032160 -0.2621453 -0.3240279 -0.2479331 -0.2852694
#>  [4,] -0.2358057 -0.2915404 -0.2790972 -0.3094666 -0.2521563 -0.2396256
#>  [5,] -0.3115943 -0.2127067 -0.3113879 -0.2086567 -0.3837769 -0.3520286
#>  [6,] -0.2973821 -0.2509043 -0.2715738 -0.3188474 -0.3978541 -0.3131655
#>  [7,] -0.3171109 -0.4066595 -0.3095373 -0.3713689 -0.3572573 -0.3488783
#>  [8,] -0.3750437 -0.4390058 -0.4296194 -0.3978895 -0.3108680 -0.3144877
#>  [9,] -0.3798103 -0.3108303 -0.4285184 -0.2780777 -0.3573579 -0.4944807
#> [10,] -0.2767894 -0.2852693 -0.2844605 -0.3431232 -0.2919698 -0.3350639
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3034022 -0.3884209 -0.4464434 -0.4476836 -0.1498976 -0.3772800
#>  [2,] -0.2831185 -0.3702982 -0.3645083 -0.2616748 -0.3128512 -0.3912174
#>  [3,] -0.2469726 -0.2134628 -0.2126626 -0.2523679 -0.2272547 -0.2079660
#>  [4,] -0.2509937 -0.2679290 -0.2828654 -0.3009236 -0.2893532 -0.3459523
#>  [5,] -0.2752244 -0.2929550 -0.3397776 -0.3812315 -0.1977668 -0.3936415
#>  [6,] -0.4716555 -0.4124588 -0.2994377 -0.3471654 -0.2279799 -0.2592685
#>  [7,] -0.4633510 -0.4228969 -0.3988834 -0.4466372 -0.1317509 -0.3819985
#>  [8,] -0.4199422 -0.3480206 -0.4080041 -0.2847540 -0.1551439 -0.3667463
#>  [9,] -0.3082256 -0.3290349 -0.3026527 -0.3181538 -0.1507632 -0.4322966
#> [10,] -0.3523229 -0.3480555 -0.3024336 -0.3825877 -0.1413767 -0.3223045
#>               V7         V8          V9
#>  [1,] -0.2574235 -0.1055416 -0.10045429
#>  [2,] -0.2200516 -0.2299726 -0.12372323
#>  [3,] -0.1871153 -0.1374213 -0.19011271
#>  [4,] -0.1026027 -0.1852292 -0.05596114
#>  [5,] -0.2764640 -0.2813377 -0.27957253
#>  [6,] -0.2401641 -0.2885812 -0.22955073
#>  [7,] -0.1493852 -0.1480067 -0.17816673
#>  [8,] -0.1637509 -0.2659470 -0.16128589
#>  [9,] -0.2549100 -0.1030988 -0.05298184
#> [10,] -0.2276210 -0.1331397 -0.10323261
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]       [,4]         [,5]       [,6]
#> [1,] 0.071862579  0.07798582  0.097487847 0.02010323 -0.025619265  0.1709019
#> [2,] 0.008574541 -0.05236113 -0.006261447 0.03958917 -0.008374195 -0.1556539
#>            [,7]        [,8]        [,9]      [,10]
#> [1,]  0.1232264 -0.05230932 -0.02872934  0.0798191
#> [2,] -0.1801090  0.07087848 -0.12614702 -0.1857559
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.208514e-04  0.0014173603  0.0017072777  0.0017018439  0.0015468562
#>  [2,]  5.642971e-05  0.0003461585  0.0004104114  0.0004061913  0.0004174888
#>  [3,] -4.283040e-05 -0.0003304763 -0.0004180188 -0.0004380147 -0.0003212677
#>  [4,]  1.745524e-04  0.0012601659  0.0015617191  0.0015886827  0.0014013819
#>  [5,]  1.529833e-04  0.0011245093  0.0014038545  0.0014403447  0.0012758429
#>  [6,] -3.184932e-04 -0.0020609305 -0.0024774185 -0.0024236663 -0.0020884027
#>  [7,] -1.904621e-04 -0.0011484586 -0.0013705744 -0.0013172003 -0.0011434475
#>  [8,]  6.047662e-04  0.0040344195  0.0048686923  0.0048609989  0.0043615467
#>  [9,]  8.101994e-05  0.0005438072  0.0006594815  0.0006600462  0.0006206839
#> [10,] -1.476654e-04 -0.0009329233 -0.0011188556 -0.0010836928 -0.0009178615
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0013589792  1.084534e-03  1.094753e-03  1.447507e-03  0.0016864613
#>  [2,]  0.0004428734  3.906855e-04  4.595416e-04  6.039317e-04  0.0006664399
#>  [3,] -0.0001515371 -5.013483e-05  4.248397e-05  2.538187e-05 -0.0000351570
#>  [4,]  0.0011467562  8.695201e-04  8.579006e-04  1.202132e-03  0.0014919306
#>  [5,]  0.0010399772  7.743098e-04  7.779905e-04  1.125361e-03  0.0013889682
#>  [6,] -0.0016676329 -1.230867e-03 -1.090696e-03 -1.441791e-03 -0.0017182596
#>  [7,] -0.0009531518 -7.728600e-04 -6.828603e-04 -8.328644e-04 -0.0009489485
#>  [8,]  0.0037257193  2.944862e-03  2.895740e-03  3.784994e-03  0.0044604841
#>  [9,]  0.0005735813  4.669803e-04  4.969220e-04  6.695026e-04  0.0007754650
#> [10,] -0.0007195150 -5.427325e-04 -4.562666e-04 -5.746792e-04 -0.0006864707
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0023073735  2.590502e-04  0.0028285904  0.0029375465  0.0029150780
#>  [2,]  0.0007470663  7.770235e-05  0.0007759592  0.0007831677  0.0008444404
#>  [3,] -0.0002752033 -3.554040e-05 -0.0004945219 -0.0005386921 -0.0004682302
#>  [4,]  0.0020717377  2.031697e-04  0.0025207497  0.0026651573  0.0026778447
#>  [5,]  0.0018801593  1.780766e-04  0.0022344599  0.0023741206  0.0024391699
#>  [6,] -0.0027347492 -3.520973e-04 -0.0037229148 -0.0038942003 -0.0036829676
#>  [7,] -0.0015146265 -2.082239e-04 -0.0021149915 -0.0022039862 -0.0020258944
#>  [8,]  0.0063145516  6.948529e-04  0.0079629636  0.0083528137  0.0082245395
#>  [9,]  0.0009781091  9.890630e-05  0.0011318429  0.0011750944  0.0012185261
#> [10,] -0.0011643709 -1.594383e-04 -0.0016577328 -0.0017415324 -0.0015989626
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0028582255  0.0030035523  0.0029877637  0.0027070885  0.0026914029
#>  [2,]  0.0009391744  0.0010472589  0.0010129273  0.0008865575  0.0008668107
#>  [3,] -0.0003041691 -0.0002278858 -0.0002647951 -0.0002437665 -0.0002523567
#>  [4,]  0.0025914644  0.0027214223  0.0026797950  0.0023527504  0.0022996344
#>  [5,]  0.0024227895  0.0025621083  0.0025290561  0.0022266300  0.0022005896
#>  [6,] -0.0033196654 -0.0033621118 -0.0034504130 -0.0031273178 -0.0030903592
#>  [7,] -0.0017633750 -0.0017674409 -0.0017975742 -0.0016169604 -0.0015808129
#>  [8,]  0.0078821873  0.0082200033  0.0082419593  0.0075311145  0.0074596857
#>  [9,]  0.0012544651  0.0013452569  0.0013240162  0.0011876085  0.0011714458
#> [10,] -0.0013787927 -0.0013777814 -0.0014205229 -0.0012925076 -0.0012776582
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0027244771  0.0022075997  2.900357e-04  0.0018226496  0.0014506774
#>  [2,]  0.0008729644  0.0007764808  8.098298e-05  0.0007006280  0.0006568806
#>  [3,] -0.0002505803 -0.0000592731 -4.811327e-05  0.0000690565  0.0001982681
#>  [4,]  0.0022760467  0.0017855049  2.287855e-04  0.0013515690  0.0009641571
#>  [5,]  0.0021744384  0.0017471061  2.030473e-04  0.0013346994  0.0009931043
#>  [6,] -0.0031361054 -0.0023041923 -4.000780e-04 -0.0018017987 -0.0012205539
#>  [7,] -0.0016270045 -0.0011629583 -2.366864e-04 -0.0009289426 -0.0005972570
#>  [8,]  0.0074912408  0.0059345358  7.851344e-04  0.0047920583  0.0036439666
#>  [9,]  0.0011585221  0.0009724611  1.094573e-04  0.0008190531  0.0006899773
#> [10,] -0.0013107152 -0.0009255583 -1.816796e-04 -0.0007119257 -0.0004267242
#>                 V32           V33           V34           V35          V36
#>  [1,]  0.0016287103  0.0013413585  0.0011312602  7.349452e-04 4.347007e-04
#>  [2,]  0.0006876556  0.0006397423  0.0005926996  5.295574e-04 5.379216e-04
#>  [3,]  0.0001202008  0.0002412685  0.0003359198  4.951639e-04 6.563074e-04
#>  [4,]  0.0011501659  0.0009051727  0.0006890702  2.930526e-04 3.772529e-05
#>  [5,]  0.0011393271  0.0009045894  0.0006898459  3.237016e-04 1.193157e-04
#>  [6,] -0.0015938269 -0.0010834461 -0.0007224915 -5.527366e-05 5.647142e-04
#>  [7,] -0.0007874444 -0.0005316299 -0.0004030029 -8.587026e-05 3.010606e-04
#>  [8,]  0.0041590329  0.0033288861  0.0027238620  1.579481e-03 6.964923e-04
#>  [9,]  0.0007601507  0.0006676402  0.0005854845  4.432767e-04 3.716717e-04
#> [10,] -0.0005851514 -0.0003554221 -0.0002162125  7.372032e-05 3.752951e-04
#>                V37           V38           V39            V4           V40
#>  [1,] 3.887442e-04  0.0009591986  1.299549e-03  4.522331e-04  0.0008405754
#>  [2,] 4.659303e-04  0.0004942279  4.973703e-04  1.141891e-04  0.0004082505
#>  [3,] 5.456022e-04  0.0002205576 -7.372469e-06 -9.675728e-05  0.0001180350
#>  [4,] 7.413551e-05  0.0006166888  9.819578e-04  3.668374e-04  0.0006357345
#>  [5,] 1.499599e-04  0.0006132677  9.261701e-04  3.199642e-04  0.0006366196
#>  [6,] 4.604152e-04 -0.0006907266 -1.390246e-03 -6.681866e-04 -0.0006272802
#>  [7,] 2.949847e-04 -0.0003456068 -7.144866e-04 -3.908032e-04 -0.0002696516
#>  [8,] 6.510185e-04  0.0023618608  3.442761e-03  1.235482e-03  0.0021412558
#>  [9,] 3.241121e-04  0.0004803149  5.745268e-04  1.666176e-04  0.0004214238
#> [10,] 3.180686e-04 -0.0002249958 -5.473381e-04 -3.052604e-04 -0.0001868436
#>                 V41           V42           V43           V44           V45
#>  [1,]  8.815981e-04  0.0013167020  0.0014586509  0.0015642691  0.0017025408
#>  [2,]  3.515838e-04  0.0004133726  0.0004186477  0.0004640190  0.0004595139
#>  [3,]  1.688841e-05 -0.0001671149 -0.0002435303 -0.0002420202 -0.0003304062
#>  [4,]  6.205847e-04  0.0010286728  0.0012097290  0.0013054090  0.0014005318
#>  [5,]  6.343083e-04  0.0009887261  0.0011309048  0.0012225981  0.0012879953
#>  [6,] -8.916858e-04 -0.0016551429 -0.0019443649 -0.0021120824 -0.0024776394
#>  [7,] -4.056962e-04 -0.0008315377 -0.0010264740 -0.0010950492 -0.0013102453
#>  [8,]  2.258758e-03  0.0035598556  0.0040500298  0.0043041579  0.0047149790
#>  [9,]  4.025607e-04  0.0005451490  0.0005879862  0.0006441034  0.0006689194
#> [10,] -3.150264e-04 -0.0006735548 -0.0008359749 -0.0009001178 -0.0010770152
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0013600087  0.0008687079  0.0006228277  3.108776e-04  5.181476e-04
#>  [2,]  0.0003711131  0.0002414948  0.0001829907  8.624129e-05  1.463032e-04
#>  [3,] -0.0002582408 -0.0001642840 -0.0001056549 -6.207876e-05 -9.715799e-05
#>  [4,]  0.0011059730  0.0007220675  0.0005084497  2.581949e-04  4.310769e-04
#>  [5,]  0.0010217763  0.0006813901  0.0004771938  2.425306e-04  3.799076e-04
#>  [6,] -0.0019613564 -0.0012148299 -0.0008601031 -4.419536e-04 -7.144524e-04
#>  [7,] -0.0010363677 -0.0006284933 -0.0004503021 -2.296679e-04 -4.092597e-04
#>  [8,]  0.0037218628  0.0023598772  0.0016933033  8.581022e-04  1.415221e-03
#>  [9,]  0.0005245748  0.0003362154  0.0002447513  1.201008e-04  1.998259e-04
#> [10,] -0.0008495944 -0.0005190600 -0.0003675329 -1.903822e-04 -3.169848e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  8.631494e-05  1.221425e-04  1.129290e-04  5.568728e-05  5.412373e-05
#>  [2,]  2.801877e-05  3.453897e-05  3.213573e-05  1.684658e-05  1.543614e-05
#>  [3,] -8.678034e-06 -2.057455e-05 -1.884502e-05 -6.387796e-06 -7.433194e-06
#>  [4,]  6.521254e-05  9.922792e-05  8.909582e-05  4.185515e-05  4.066125e-05
#>  [5,]  6.238355e-05  9.234703e-05  8.026998e-05  3.823481e-05  3.667850e-05
#>  [6,] -1.103724e-04 -1.674076e-04 -1.570281e-04 -7.082883e-05 -7.067605e-05
#>  [7,] -5.761122e-05 -9.030753e-05 -8.756069e-05 -3.996671e-05 -4.080481e-05
#>  [8,]  2.306679e-04  3.311117e-04  3.052389e-04  1.487702e-04  1.481571e-04
#>  [9,]  3.533268e-05  4.768341e-05  4.389950e-05  2.183871e-05  2.116374e-05
#> [10,] -4.590246e-05 -7.271937e-05 -6.887391e-05 -3.056802e-05 -3.129685e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  5.307580e-05  3.835207e-05  3.077970e-05  4.618432e-05  4.526028e-05
#>  [2,]  2.047942e-05  1.160857e-05  1.222820e-05  1.190691e-05  1.300526e-05
#>  [3,] -1.142032e-07 -4.428600e-06  1.759451e-06 -8.047508e-06 -6.092788e-06
#>  [4,]  3.715949e-05  2.919850e-05  1.927762e-05  3.510858e-05  3.203979e-05
#>  [5,]  3.431680e-05  2.595629e-05  1.803222e-05  3.097885e-05  2.806641e-05
#>  [6,] -6.240187e-05 -4.994976e-05 -3.278256e-05 -6.369107e-05 -5.973054e-05
#>  [7,] -3.385168e-05 -2.855744e-05 -1.903290e-05 -3.729039e-05 -3.540258e-05
#>  [8,]  1.364507e-04  1.043551e-04  7.817310e-05  1.271537e-04  1.212490e-04
#>  [9,]  2.285189e-05  1.514345e-05  1.332964e-05  1.701570e-05  1.710026e-05
#> [10,] -2.507071e-05 -2.175929e-05 -1.354186e-05 -2.883172e-05 -2.648790e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  4.151245e-04  4.049437e-05  5.123627e-04  0.0006542502  0.0011592788
#>  [2,]  1.295414e-04  1.322454e-05  1.762772e-04  0.0001727551  0.0002632445
#>  [3,] -4.970059e-05 -3.733676e-06 -4.271355e-05 -0.0001244697 -0.0002815749
#>  [4,]  3.364288e-04  2.980695e-05  4.365111e-04  0.0005711450  0.0009955750
#>  [5,]  3.065076e-04  2.606455e-05  3.986819e-04  0.0005095145  0.0008806111
#>  [6,] -5.070851e-04 -5.207021e-05 -6.059277e-04 -0.0009041360 -0.0017425809
#>  [7,] -2.912723e-04 -3.060214e-05 -3.418479e-04 -0.0005170575 -0.0009933350
#>  [8,]  1.142541e-03  1.069277e-04  1.394372e-03  0.0018701337  0.0033185499
#>  [9,]  1.665592e-04  1.609953e-05  2.172670e-04  0.0002550242  0.0004321930
#> [10,] -2.237919e-04 -2.286805e-05 -2.603516e-04 -0.0004119930 -0.0008034724
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.0860942 -0.08568292 -0.04742931 -0.03696952 -0.03515725 -0.09179605
#> [2,]  0.0860942  0.08568292  0.04742931  0.03696952  0.03515725  0.09179605
#>            [,7]        [,8]        [,9]       [,10]
#> [1,] -0.1035522 -0.07156528 -0.06955875 -0.09914753
#> [2,]  0.1035522  0.07156528  0.06955875  0.09914753
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6850414 0.5867398 0.2867409 0.1225872 0.3288780 0.4107096 0.6332745
#>  [8] 0.5083497 0.5182180 0.5453364
#> 
#> $B[[2]]
#> [1]  0.09472697 -0.16904634
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  3.367984e-03  1.267177e-03 -1.867152e-05  2.677054e-03  2.527769e-03
#>  [6] -3.444346e-03 -1.863111e-03  8.997683e-03  1.503113e-03 -1.387326e-03
#> 
#> $vB[[2]]
#> [1] -0.1315242  0.1315242
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#>  [2,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#>  [3,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#>  [4,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#>  [5,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#>  [6,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#>  [7,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#>  [8,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#>  [9,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [10,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [11,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [12,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [13,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [14,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [15,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [16,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [17,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [18,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [19,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [20,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [21,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [22,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [23,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [24,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [25,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [26,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [27,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [28,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [29,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [30,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [31,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [32,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [33,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [34,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [35,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [36,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [37,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [38,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [39,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [40,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#>  [2,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#>  [3,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#>  [4,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#>  [5,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#>  [6,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#>  [7,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#>  [8,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#>  [9,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [10,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [11,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [12,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [13,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [14,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [15,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [16,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [17,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [18,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [19,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [20,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [21,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [22,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [23,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [24,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [25,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [26,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [27,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [28,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [29,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [30,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [31,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [32,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [33,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [34,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [35,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [36,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [37,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [38,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [39,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [40,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#>  [2,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#>  [3,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#>  [4,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#>  [5,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#>  [6,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#>  [7,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#>  [8,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#>  [9,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [10,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [11,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [12,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [13,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [14,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [15,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [16,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [17,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [18,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [19,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [20,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [21,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [22,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [23,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [24,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [25,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [26,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [27,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [28,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [29,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [30,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [31,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [32,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [33,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [34,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [35,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [36,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [37,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [38,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [39,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [40,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#>  [2,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#>  [3,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#>  [4,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#>  [5,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#>  [6,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#>  [7,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#>  [8,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#>  [9,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [10,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [11,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [12,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [13,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [14,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [15,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [16,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [17,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [18,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [19,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [20,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [21,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [22,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [23,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [24,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [25,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [26,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [27,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [28,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [29,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [30,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [31,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [32,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [33,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [34,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [35,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [36,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [37,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [38,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [39,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [40,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#>  [2,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#>  [3,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#>  [4,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#>  [5,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#>  [6,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#>  [7,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#>  [8,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#>  [9,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [10,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [11,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [12,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [13,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [14,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [15,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [16,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [17,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [18,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [19,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [20,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [21,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [22,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [23,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [24,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [25,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [26,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [27,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [28,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [29,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [30,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [31,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [32,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [33,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [34,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [35,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [36,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [37,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [38,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [39,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [40,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#>  [2,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#>  [3,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#>  [4,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#>  [5,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#>  [6,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#>  [7,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#>  [8,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#>  [9,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [10,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [11,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [12,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [13,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [14,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [15,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [16,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [17,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [18,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [19,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [20,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [21,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [22,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [23,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [24,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [25,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [26,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [27,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [28,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [29,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [30,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [31,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [32,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [33,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [34,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [35,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [36,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [37,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [38,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [39,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [40,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7495661 0.7707731 0.5376688 0.4168170 0.4116709 0.7744572 0.8474120
#>  [2,] 0.6913246 0.7041533 0.4098706 0.3402548 0.3004473 0.5396921 0.7143639
#>  [3,] 0.8061915 0.8559265 0.5929614 0.4290390 0.4930864 0.8176511 0.8980672
#>  [4,] 0.7377501 0.7851222 0.5207958 0.3699512 0.4340010 0.7369401 0.8380136
#>  [5,] 0.7262907 0.7770243 0.5644631 0.4463134 0.5575662 0.7213528 0.8266889
#>  [6,] 0.7204750 0.7636906 0.5225405 0.3865594 0.4506267 0.7259345 0.8429176
#>  [7,] 0.7304597 0.7561643 0.4284783 0.3329189 0.3087652 0.7482291 0.8196019
#>  [8,] 0.7495650 0.7811418 0.5486888 0.4191351 0.4654943 0.7947138 0.8687741
#>  [9,] 0.7734046 0.8138801 0.5761008 0.4320137 0.5006953 0.8196832 0.8835359
#> [10,] 0.7150596 0.7284326 0.4136667 0.3082833 0.3123829 0.7092074 0.8236730
#> [11,] 0.6984008 0.7551914 0.3925240 0.2524081 0.2870713 0.5860112 0.7986872
#> [12,] 0.6768981 0.6349733 0.3295990 0.2937517 0.1983581 0.5380951 0.6860892
#> [13,] 0.7144727 0.7239882 0.4330823 0.3865997 0.3302266 0.7301623 0.7888282
#> [14,] 0.7853002 0.8291701 0.5417342 0.4170697 0.4528673 0.8029839 0.8751333
#> [15,] 0.7315783 0.7342901 0.5115607 0.4243543 0.4467237 0.7603842 0.8221843
#> [16,] 0.7375778 0.7298651 0.5203885 0.4238513 0.3921637 0.6631396 0.7818706
#> [17,] 0.7658414 0.8042386 0.5753535 0.4253283 0.5047379 0.7717915 0.8637789
#> [18,] 0.7044682 0.7180496 0.4052333 0.3060536 0.3145685 0.6981014 0.8099807
#> [19,] 0.7082318 0.7409604 0.3936241 0.2776075 0.2833296 0.5925433 0.7892142
#> [20,] 0.7266171 0.7524891 0.4457440 0.3579712 0.3721472 0.7636794 0.8260300
#> [21,] 0.6780783 0.6842333 0.3827654 0.3381641 0.3259514 0.6298820 0.7859744
#> [22,] 0.7112643 0.7310181 0.4273648 0.3335406 0.3328941 0.7345504 0.8229411
#> [23,] 0.7628916 0.8179495 0.4839780 0.3529003 0.3757643 0.7585020 0.8556788
#> [24,] 0.6763887 0.6241015 0.3894529 0.4222570 0.3038165 0.6970074 0.7245101
#> [25,] 0.7504269 0.7754816 0.5254702 0.4198418 0.4203268 0.8059568 0.8717025
#> [26,] 0.7511241 0.7866878 0.5378853 0.4025231 0.4538734 0.7977633 0.8702646
#> [27,] 0.6352578 0.6563879 0.3639391 0.3009562 0.2592857 0.5382391 0.6950845
#> [28,] 0.7222610 0.7279582 0.4782673 0.4166379 0.4186836 0.7153646 0.7766675
#> [29,] 0.6884007 0.7425817 0.4563977 0.3724580 0.3252554 0.6757887 0.8069172
#> [30,] 0.6736214 0.7320285 0.3854986 0.2546142 0.2924102 0.6055903 0.8040045
#> [31,] 0.5585717 0.5742562 0.3100396 0.2216357 0.2337064 0.4629473 0.6956598
#> [32,] 0.6118070 0.5486467 0.3475379 0.3934340 0.2865763 0.6259197 0.6658014
#> [33,] 0.6475345 0.5746636 0.2686957 0.2401277 0.1584665 0.4878377 0.6644299
#> [34,] 0.7542308 0.7953529 0.5329197 0.3616237 0.4261929 0.7123592 0.8530719
#> [35,] 0.5622915 0.5565898 0.3360755 0.2979916 0.2581487 0.5073990 0.6312170
#> [36,] 0.7103273 0.7604520 0.4255355 0.3552038 0.3867076 0.6740088 0.8097793
#> [37,] 0.7659179 0.8098197 0.5759753 0.4268707 0.5023640 0.8092428 0.8799161
#> [38,] 0.6325944 0.6426183 0.3545629 0.2652400 0.2712405 0.6205311 0.7650706
#> [39,] 0.7704174 0.8114110 0.5850597 0.4332456 0.5053147 0.7727796 0.8693343
#> [40,] 0.7003104 0.7202911 0.4259263 0.3431126 0.3184247 0.5506517 0.7404284
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6743558 0.6717691 0.8154957
#>  [2,] 0.6416885 0.4511746 0.6600165
#>  [3,] 0.7714097 0.7325796 0.8487759
#>  [4,] 0.6945871 0.6423594 0.7965118
#>  [5,] 0.7360893 0.7038103 0.7686483
#>  [6,] 0.7085589 0.6683678 0.7964883
#>  [7,] 0.6397528 0.5962108 0.7898941
#>  [8,] 0.6936584 0.7187360 0.8285147
#>  [9,] 0.7308329 0.7361826 0.8495084
#> [10,] 0.6242785 0.5612449 0.7755601
#> [11,] 0.6681134 0.4884535 0.7115046
#> [12,] 0.5727182 0.3913143 0.6438109
#> [13,] 0.6303119 0.5781130 0.7794732
#> [14,] 0.7367909 0.6958110 0.8266070
#> [15,] 0.6728937 0.6605102 0.7899413
#> [16,] 0.6808201 0.5930924 0.7416227
#> [17,] 0.7438256 0.6955101 0.8172233
#> [18,] 0.6216754 0.5498156 0.7629295
#> [19,] 0.6647678 0.4877159 0.7098689
#> [20,] 0.6386789 0.6271489 0.7784183
#> [21,] 0.6448764 0.5663257 0.7294077
#> [22,] 0.6139054 0.5903983 0.7849079
#> [23,] 0.7098337 0.6228806 0.7933124
#> [24,] 0.5709536 0.5387697 0.7328292
#> [25,] 0.6931140 0.7145608 0.8371345
#> [26,] 0.6965721 0.7127787 0.8337361
#> [27,] 0.5804071 0.4455110 0.6401226
#> [28,] 0.6629074 0.6143161 0.7390068
#> [29,] 0.6308383 0.6072440 0.7503900
#> [30,] 0.6400776 0.4722789 0.7421996
#> [31,] 0.5209427 0.3586537 0.6168347
#> [32,] 0.5218408 0.4956312 0.6714575
#> [33,] 0.4866893 0.3204880 0.6069899
#> [34,] 0.7333927 0.6243276 0.7996299
#> [35,] 0.5044666 0.4498609 0.5981742
#> [36,] 0.6810999 0.6142298 0.7444191
#> [37,] 0.7251474 0.7241669 0.8400150
#> [38,] 0.5807614 0.4634045 0.7214410
#> [39,] 0.7450050 0.7099615 0.8189821
#> [40,] 0.6534041 0.4866739 0.6670405
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4642613 0.5357387
#>  [2,] 0.4607567 0.5392433
#>  [3,] 0.4561798 0.5438202
#>  [4,] 0.4620893 0.5379107
#>  [5,] 0.4534533 0.5465467
#>  [6,] 0.4605635 0.5394365
#>  [7,] 0.4699305 0.5300695
#>  [8,] 0.4626380 0.5373620
#>  [9,] 0.4596050 0.5403950
#> [10,] 0.4709508 0.5290492
#> [11,] 0.4647267 0.5352733
#> [12,] 0.4691658 0.5308342
#> [13,] 0.4682447 0.5317553
#> [14,] 0.4592861 0.5407139
#> [15,] 0.4628066 0.5371934
#> [16,] 0.4588556 0.5411444
#> [17,] 0.4568243 0.5431757
#> [18,] 0.4706857 0.5293143
#> [19,] 0.4643113 0.5356887
#> [20,] 0.4684072 0.5315928
#> [21,] 0.4654270 0.5345730
#> [22,] 0.4714496 0.5285504
#> [23,] 0.4629841 0.5370159
#> [24,] 0.4707900 0.5292100
#> [25,] 0.4642814 0.5357186
#> [26,] 0.4632544 0.5367456
#> [27,] 0.4677425 0.5322575
#> [28,] 0.4614387 0.5385613
#> [29,] 0.4669243 0.5330757
#> [30,] 0.4688159 0.5311841
#> [31,] 0.4747377 0.5252623
#> [32,] 0.4730066 0.5269934
#> [33,] 0.4760826 0.5239174
#> [34,] 0.4590873 0.5409127
#> [35,] 0.4724341 0.5275659
#> [36,] 0.4616460 0.5383540
#> [37,] 0.4598163 0.5401837
#> [38,] 0.4742683 0.5257317
#> [39,] 0.4564226 0.5435774
#> [40,] 0.4599141 0.5400859
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]       [,4]         [,5]        [,6]
#>  [1,] 1.0962995 1.2126817  0.15096121 -0.3358537 -0.357062365  1.23365271
#>  [2,] 0.8063191 0.8671548 -0.36450023 -0.6621589 -0.845168783  0.15910305
#>  [3,] 1.4254508 1.7818610  0.37622142 -0.2857732 -0.027656016  1.50051373
#>  [4,] 1.0343074 1.2957697  0.08323108 -0.5324262 -0.265545371  1.03012475
#>  [5,] 0.9758838 1.2484087  0.25929536 -0.2155773  0.231290452  0.95118182
#>  [6,] 0.9468192 1.1730210  0.09022324 -0.4617977 -0.198138945  0.97409291
#>  [7,] 0.9969559 1.1317642 -0.28806245 -0.6950129 -0.805898378  1.08918957
#>  [8,] 1.0962936 1.2723324  0.19537419 -0.3263247 -0.138242561  1.35357714
#>  [9,] 1.2276361 1.4754223  0.30678713 -0.2736399  0.002781021  1.51420278
#> [10,] 0.9200861 0.9866848 -0.34882771 -0.8081574 -0.789002236  0.89153794
#> [11,] 0.8396944 1.1264946 -0.43671502 -1.0858103 -0.909650911  0.34749979
#> [12,] 0.7395528 0.5536124 -0.70999908 -0.8772319 -1.396588075  0.15267633
#> [13,] 0.9172071 0.9643317 -0.26928653 -0.4616279 -0.707160535  0.99544596
#> [14,] 1.2968255 1.5797569  0.16732611 -0.3348143 -0.189092330  1.40504891
#> [15,] 1.0026449 1.0164990  0.04625096 -0.3049238 -0.213917119  1.15478713
#> [16,] 1.0334172 0.9939383  0.08159937 -0.3069832 -0.438226202  0.67731685
#> [17,] 1.1849765 1.4129992  0.30372770 -0.3009376  0.018952167  1.21845466
#> [18,] 0.8686670 0.9348077 -0.38370615 -0.8186342 -0.778846577  0.83827308
#> [19,] 0.8868118 1.0509663 -0.43210350 -0.9563605 -0.928005021  0.37448911
#> [20,] 0.9775265 1.1119320 -0.21788190 -0.5841806 -0.523016433  1.17295848
#> [21,] 0.7449543 0.7732954 -0.47782706 -0.6714865 -0.726554111  0.53171078
#> [22,] 0.9015324 0.9997939 -0.29261075 -0.6922145 -0.695124281  1.01783355
#> [23,] 1.1685983 1.5025167 -0.06410984 -0.6063148 -0.507566070  1.14448424
#> [24,] 0.7372248 0.5069940 -0.44961234 -0.3135149 -0.829189591  0.83308771
#> [25,] 1.1008904 1.2395268  0.10196900 -0.3234230 -0.321431954  1.42394947
#> [26,] 1.1046166 1.3050748  0.15183230 -0.3949632 -0.185032326  1.37237302
#> [27,] 0.5548401 0.6472384 -0.55830765 -0.8427486 -1.049684300  0.15325546
#> [28,] 0.9557049 0.9842879 -0.08698563 -0.3365907 -0.328179553  0.92158340
#> [29,] 0.7926528 1.0594303 -0.17485323 -0.5216865 -0.729723672  0.73448501
#> [30,] 0.7246104 1.0049390 -0.46627370 -1.0741526 -0.883706886  0.42881329
#> [31,] 0.2353673 0.2992377 -0.79993412 -1.2561595 -1.187499521 -0.14848293
#> [32,] 0.4549143 0.1952042 -0.62987950 -0.4329001 -0.912070752  0.51475156
#> [33,] 0.6082192 0.3009045 -1.00125007 -1.1519793 -1.669682493 -0.04865879
#> [34,] 1.1213054 1.3574987  0.13186939 -0.5683239 -0.297401268  0.90686981
#> [35,] 0.2504672 0.2273333 -0.68083259 -0.8568799 -1.055613176  0.02959800
#> [36,] 0.8969741 1.1551592 -0.30008993 -0.5962427 -0.461172838  0.72637259
#> [37,] 1.1854032 1.4488393  0.30627313 -0.2946301  0.009456255  1.44509765
#> [38,] 0.5433629 0.5867465 -0.59904202 -1.0189089 -0.988337706  0.49180304
#> [39,] 1.2106693 1.4592048  0.34357920 -0.2686211  0.021259781  1.22407339
#> [40,] 0.8487762 0.9459059 -0.29849136 -0.6494539 -0.761020767  0.20330424
#>            [,7]        [,8]        [,9]     [,10]
#>  [1,] 1.7144453  0.72795238  0.71619718 1.4861232
#>  [2,] 0.9166741  0.58270027 -0.19592582 0.6633677
#>  [3,] 2.1759312  1.21628833  1.00774987 1.7250326
#>  [4,] 1.6435220  0.82165313  0.58561918 1.3646337
#>  [5,] 1.5623405  1.02574076  0.86550879 1.2006943
#>  [6,] 1.6800990  0.88839541  0.70081191 1.3644891
#>  [7,] 1.5136525  0.57429141  0.38970113 1.3242874
#>  [8,] 1.8901631  0.81727890  0.93820039 1.5751372
#>  [9,] 2.0263490  0.99885230  1.02622113 1.7307508
#> [10,] 1.5414332  0.50774827  0.24621576 1.2399776
#> [11,] 1.3781097  0.69966456 -0.04619427 0.9027029
#> [12,] 0.7818986  0.29295014 -0.44179088 0.5919432
#> [13,] 1.3178767  0.53355514  0.31503187 1.2625993
#> [14,] 1.9471296  1.02935546  0.82742887 1.5617688
#> [15,] 1.5312168  0.72130222  0.66556855 1.3245719
#> [16,] 1.2766006  0.75754327  0.37676422 1.0544196
#> [17,] 1.8470372  1.06594792  0.82600783 1.4976469
#> [18,] 1.4498847  0.49666538  0.19992548 1.1688083
#> [19,] 1.3201952  0.68461442 -0.04914620 0.8947476
#> [20,] 1.5577481  0.56963482  0.52000505 1.2564724
#> [21,] 1.3008286  0.59659273  0.26687545 0.9916193
#> [22,] 1.5364024  0.46375845  0.36561222 1.2945003
#> [23,] 1.7798535  0.89457658  0.50179283 1.3450089
#> [24,] 0.9669449  0.28574280  0.15539053 1.0090246
#> [25,] 1.9160960  0.81471802  0.91763894 1.6370603
#> [26,] 1.9032999  0.83102733  0.90891778 1.6123408
#> [27,] 0.8239988  0.32444485 -0.21882499 0.5758964
#> [28,] 1.2463507  0.67627747  0.46549154 1.0408127
#> [29,] 1.4301022  0.53581468  0.43574197 1.1006931
#> [30,] 1.4115133  0.57570084 -0.11099825 1.0574326
#> [31,] 0.8267146  0.08381991 -0.58121232 0.4761345
#> [32,] 0.6892558  0.08741871 -0.01747548 0.7147844
#> [33,] 0.6830986 -0.05325544 -0.75152984 0.4346768
#> [34,] 1.7589005  1.01190476  0.50795773 1.3839826
#> [35,] 0.5374414  0.01786689 -0.20123292 0.3978635
#> [36,] 1.4485766  0.75883088  0.46512741 1.0690650
#> [37,] 1.9916355  0.97014024  0.96522572 1.6583394
#> [38,] 1.1806833  0.32589975 -0.14664411 0.9516205
#> [39,] 1.8950854  1.07214710  0.89519707 1.5094666
#> [40,] 1.0481964  0.63403643 -0.05331709 0.6948297
#> 
#> $pre[[3]]
#>             [,1]         [,2]
#>  [1,] -0.1459211 -0.002722200
#>  [2,] -0.1353811  0.021915607
#>  [3,] -0.1639981  0.011733402
#>  [4,] -0.1496761  0.002258306
#>  [5,] -0.1681608  0.018566842
#>  [6,] -0.1555292  0.002545184
#>  [7,] -0.1337476 -0.013324204
#>  [8,] -0.1539959 -0.004268860
#>  [9,] -0.1600315  0.001901313
#> [10,] -0.1312369 -0.014909059
#> [11,] -0.1370579  0.004270234
#> [12,] -0.1176147  0.005878729
#> [13,] -0.1341366 -0.006944259
#> [14,] -0.1566936  0.006523436
#> [15,] -0.1487767  0.000272389
#> [16,] -0.1469690  0.017981584
#> [17,] -0.1611850  0.011949003
#> [18,] -0.1309816 -0.013589729
#> [19,] -0.1363819  0.006615973
#> [20,] -0.1384123 -0.011872644
#> [21,] -0.1424215 -0.003908591
#> [22,] -0.1319604 -0.017634502
#> [23,] -0.1461617  0.002173273
#> [24,] -0.1257727 -0.008799526
#> [25,] -0.1511355 -0.008017324
#> [26,] -0.1532209 -0.005973149
#> [27,] -0.1253121  0.003897297
#> [28,] -0.1461226  0.008429354
#> [29,] -0.1385126 -0.006016168
#> [30,] -0.1311175 -0.006218790
#> [31,] -0.1129857 -0.011850318
#> [32,] -0.1201056 -0.012026944
#> [33,] -0.1025244 -0.006781621
#> [34,] -0.1535107  0.010506674
#> [35,] -0.1186871 -0.008311734
#> [36,] -0.1508328  0.002885274
#> [37,] -0.1587430  0.002339186
#> [38,] -0.1212494 -0.018231728
#> [39,] -0.1624011  0.012351867
#> [40,] -0.1396667  0.021021648
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5357387 -0.5357387
#>  [2,] -0.4607567  0.4607567
#>  [3,] -0.4561798  0.4561798
#>  [4,]  0.5379107 -0.5379107
#>  [5,] -0.4534533  0.4534533
#>  [6,]  0.5394365 -0.5394365
#>  [7,]  0.5300695 -0.5300695
#>  [8,]  0.5373620 -0.5373620
#>  [9,]  0.5403950 -0.5403950
#> [10,]  0.5290492 -0.5290492
#> [11,]  0.5352733 -0.5352733
#> [12,] -0.4691658  0.4691658
#> [13,]  0.5317553 -0.5317553
#> [14,]  0.5407139 -0.5407139
#> [15,] -0.4628066  0.4628066
#> [16,] -0.4588556  0.4588556
#> [17,] -0.4568243  0.4568243
#> [18,]  0.5293143 -0.5293143
#> [19,]  0.5356887 -0.5356887
#> [20,]  0.5315928 -0.5315928
#> [21,] -0.4654270  0.4654270
#> [22,]  0.5285504 -0.5285504
#> [23,]  0.5370159 -0.5370159
#> [24,]  0.5292100 -0.5292100
#> [25,]  0.5357186 -0.5357186
#> [26,]  0.5367456 -0.5367456
#> [27,]  0.5322575 -0.5322575
#> [28,] -0.4614387  0.4614387
#> [29,] -0.4669243  0.4669243
#> [30,]  0.5311841 -0.5311841
#> [31,]  0.5252623 -0.5252623
#> [32,]  0.5269934 -0.5269934
#> [33,]  0.5239174 -0.5239174
#> [34,] -0.4590873  0.4590873
#> [35,]  0.5275659 -0.5275659
#> [36,] -0.4616460  0.4616460
#> [37,]  0.5401837 -0.5401837
#> [38,]  0.5257317 -0.5257317
#> [39,] -0.4564226  0.4564226
#> [40,] -0.4599141  0.4599141
#> 
#> $L
#> [1] 0.6909666 0.6876502 0.6883381 0.6886721 0.6983493 0.7108723
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536