Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.3029696 -0.12216210 -0.13554134 -0.206870710 -0.19533701 -0.059043015
#>  [2,] -0.2488702 -0.10646415  0.01840669  0.004152769 -0.01413089 -0.058058661
#>  [3,] -0.3801853 -0.08871399 -0.13891320 -0.106172357 -0.13129012 -0.032382484
#>  [4,] -0.2687152 -0.16404884 -0.20876637 -0.025814973 -0.12326200 -0.069532295
#>  [5,] -0.3035899 -0.08799411 -0.11780108 -0.115863814 -0.13023406 -0.134310059
#>  [6,] -0.2691238 -0.13692435 -0.16488613 -0.118447099  0.01879312 -0.069907578
#>  [7,] -0.3553701 -0.08285000 -0.09980361 -0.116172768 -0.09048254 -0.161323395
#>  [8,] -0.3916400 -0.09703952  0.01172203 -0.024022422 -0.06370794  0.005437837
#>  [9,] -0.2183611 -0.18632616 -0.13632959 -0.088517938 -0.15499778 -0.018689767
#> [10,] -0.3430635 -0.09377453 -0.08243726  0.014022694  0.07272236  0.042562425
#>                V15          V16         V17          V18          V19
#>  [1,] -0.068870554 -0.133876170 -0.04920978  0.016553376 -0.056066519
#>  [2,] -0.109452955 -0.046990867 -0.12385717 -0.056324462  0.067534098
#>  [3,]  0.021111356 -0.074313858  0.07717922  0.071831855  0.045917578
#>  [4,] -0.123009022 -0.162312608 -0.16937136 -0.106933657 -0.028909121
#>  [5,] -0.026486978  0.004451412 -0.03004398 -0.019116201  0.023364941
#>  [6,] -0.017861664  0.077333224  0.06834648  0.005076945  0.158544255
#>  [7,] -0.152850721  0.032378289  0.05513522  0.107410835  0.006538359
#>  [8,] -0.003518537 -0.084014259 -0.02262405 -0.090051366  0.065058423
#>  [9,] -0.116622723 -0.125034280  0.00772970 -0.025132733 -0.026609900
#> [10,] -0.065429305  0.108091076  0.09301919  0.140612467  0.007969872
#>               V2         V20         V21         V22          V23        V24
#>  [1,] -0.2439208  0.13380221  0.04519663  0.14543905  0.152657806 0.06457612
#>  [2,] -0.2954986 -0.01759201  0.06791890 -0.01905985  0.113062848 0.06832740
#>  [3,] -0.2429858  0.03328278  0.12182232  0.09302657  0.236498792 0.23160755
#>  [4,] -0.3364519 -0.05481824 -0.05622036 -0.03588273  0.001890857 0.10903115
#>  [5,] -0.2206655  0.05663232  0.02469930  0.08413617 -0.030425456 0.03430774
#>  [6,] -0.3126251  0.06967245  0.04738204  0.21241266  0.046205674 0.14605207
#>  [7,] -0.3479424  0.10821550  0.03460975  0.15371873  0.028028358 0.08412928
#>  [8,] -0.2297578  0.04455919  0.06423960  0.21712436  0.109003704 0.25308464
#>  [9,] -0.1927264  0.04169093  0.21333110  0.08467281  0.089062421 0.07597584
#> [10,] -0.3476346  0.14323898  0.18724220  0.07120942  0.126624889 0.00798985
#>              V25          V26         V27         V28         V29         V3
#>  [1,] 0.06009732 -0.005420259  0.00823843 -0.01987797 -0.01835731 -0.2519970
#>  [2,] 0.09111228  0.218237394  0.19025722  0.18078688  0.16014136 -0.2594547
#>  [3,] 0.02857773  0.182412932  0.20784851  0.19786959  0.02832311 -0.3545745
#>  [4,] 0.17842890  0.183517591  0.14580982  0.14313668  0.07242063 -0.3014416
#>  [5,] 0.20786349  0.252272620  0.23367742  0.29201563  0.27116557 -0.2503949
#>  [6,] 0.19443476  0.147993801 -0.04071775 -0.01799296 -0.01444432 -0.3407383
#>  [7,] 0.19729277  0.136578763  0.23491615  0.06234463  0.10833904 -0.2029168
#>  [8,] 0.26527262  0.281307704  0.32650304  0.20980213  0.23903651 -0.2697452
#>  [9,] 0.19303543  0.111324387  0.12355146  0.08950433  0.17295688 -0.3296554
#> [10,] 0.03105066 -0.013792784  0.02219527  0.19568380  0.04336273 -0.3235480
#>                V30         V31          V32          V33          V34
#>  [1,] -0.055219307 -0.16099635  0.038374282 -0.069387606  0.027295169
#>  [2,]  0.137445348  0.03885810 -0.030867923  0.091347764  0.111719344
#>  [3,]  0.085368921  0.15964882  0.030944476  0.034223053  0.049862072
#>  [4,]  0.148837294  0.10437713  0.008616587 -0.095770889 -0.068551322
#>  [5,]  0.034638468 -0.01685342  0.081909274 -0.081877412 -0.098059081
#>  [6,]  0.034324087  0.06262159 -0.116862847  0.005263098 -0.033404355
#>  [7,]  0.130942633  0.05264608 -0.071051912  0.055794078 -0.029329014
#>  [8,]  0.069248388  0.11711797  0.110177488 -0.001937480 -0.007581941
#>  [9,]  0.009859923  0.05819656  0.025504886 -0.032594477  0.031843973
#> [10,]  0.074778284  0.06007682 -0.023482004  0.023676375 -0.053493639
#>                V35          V36          V37         V38         V39         V4
#>  [1,] -0.063879145  0.063699441 -0.006906879 -0.08279947 -0.19301626 -0.2759029
#>  [2,] -0.004871565  0.077250423 -0.044161765  0.03250827  0.02326332 -0.2775300
#>  [3,]  0.069970268 -0.024274811  0.055223686  0.08276570 -0.09421445 -0.1880762
#>  [4,]  0.045277254 -0.004773864 -0.162363601 -0.16281791 -0.11378428 -0.2586641
#>  [5,] -0.059221704 -0.138399937 -0.105243608 -0.08774700 -0.16806966 -0.3335587
#>  [6,] -0.010769113  0.012712879 -0.037067683 -0.19947906 -0.05006670 -0.3281153
#>  [7,] -0.107489927 -0.007287584  0.080895931 -0.09980611 -0.12286964 -0.2064529
#>  [8,] -0.070810377 -0.117400847 -0.142258898 -0.21489919 -0.13597837 -0.2066460
#>  [9,] -0.076065178 -0.021270138 -0.175167664 -0.09510231 -0.08094818 -0.3006951
#> [10,] -0.023112215  0.052369791  0.074039575  0.06867743  0.04994168 -0.3349336
#>               V40         V41          V42         V43         V44         V45
#>  [1,] -0.20245526 -0.20374723 -0.234003628 -0.08733768 -0.10092130 -0.14576540
#>  [2,] -0.13097303  0.05476054 -0.006560694 -0.06098742 -0.01848857 -0.07429961
#>  [3,]  0.05747778 -0.07648864 -0.045701431 -0.10219402 -0.04900455 -0.21679139
#>  [4,]  0.02297393 -0.06750603  0.009890933 -0.07465149 -0.07330525 -0.06815482
#>  [5,] -0.18626985 -0.04572535 -0.067940690 -0.05877368 -0.10209586 -0.18037895
#>  [6,] -0.13639080 -0.04312840 -0.076825376 -0.04942390 -0.15372300 -0.07928499
#>  [7,] -0.05929928 -0.11335541 -0.019243829 -0.12770958 -0.17771140 -0.23878682
#>  [8,] -0.03256131 -0.16702876 -0.118803943 -0.17736983 -0.07323709 -0.07899278
#>  [9,] -0.20871290 -0.16336828 -0.170881496 -0.14924242 -0.18999367 -0.04272627
#> [10,]  0.05044723 -0.05483494 -0.111921202 -0.11557962 -0.13984699 -0.17000641
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.27544409 -0.1081126 -0.2210027 -0.1794724 -0.2232865 -0.2800763
#>  [2,] -0.21051886 -0.1288056 -0.1839669 -0.1900064 -0.3241873 -0.2441055
#>  [3,] -0.14782741 -0.1055613 -0.1747029 -0.3599563 -0.1686110 -0.4193576
#>  [4,] -0.14252758 -0.2516928 -0.2224690 -0.3445493 -0.2853049 -0.3051192
#>  [5,] -0.13131005 -0.2243646 -0.1868980 -0.2522207 -0.2581477 -0.4178038
#>  [6,] -0.13358644 -0.3276036 -0.2425986 -0.3455366 -0.2863170 -0.3943311
#>  [7,] -0.08304205 -0.2265498 -0.1337146 -0.3832605 -0.1651520 -0.4119380
#>  [8,] -0.20664676 -0.1751581 -0.2516944 -0.2375776 -0.2096659 -0.2390056
#>  [9,] -0.17187078 -0.1446841 -0.2630456 -0.2129197 -0.1506710 -0.2118039
#> [10,] -0.01932776 -0.1218114 -0.2192054 -0.1934868 -0.2336071 -0.2966285
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2881230 -0.2533665 -0.2271904 -0.2845895 -0.3221473 -0.2522365
#>  [2,] -0.2503653 -0.3001612 -0.2803131 -0.2576421 -0.3952227 -0.3257950
#>  [3,] -0.4628495 -0.3931759 -0.2560785 -0.2661018 -0.3127948 -0.4612285
#>  [4,] -0.3233311 -0.2860714 -0.3062534 -0.2685542 -0.3322910 -0.2538190
#>  [5,] -0.3014236 -0.3170041 -0.3235962 -0.3637058 -0.3662719 -0.4020657
#>  [6,] -0.2905291 -0.2936117 -0.3666534 -0.3125258 -0.3200784 -0.4146145
#>  [7,] -0.3202164 -0.4091714 -0.3954443 -0.4284002 -0.2796288 -0.3172957
#>  [8,] -0.2623062 -0.3212030 -0.2643175 -0.4240095 -0.3554033 -0.3155672
#>  [9,] -0.3188779 -0.2293691 -0.4026479 -0.3171064 -0.2449746 -0.2853119
#> [10,] -0.3631032 -0.3666549 -0.3608011 -0.3196107 -0.2454426 -0.3264072
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.3729594 -0.3079950 -0.2705804 -0.1355321 -0.3459502 -0.23142775
#>  [2,] -0.4154940 -0.3453368 -0.3565955 -0.2687550 -0.3717754 -0.17323577
#>  [3,] -0.4483008 -0.3361990 -0.2784900 -0.2375955 -0.2798831 -0.19197576
#>  [4,] -0.3681033 -0.2741282 -0.2025015 -0.1173666 -0.3961057 -0.16060768
#>  [5,] -0.3103844 -0.3298632 -0.4107392 -0.1114184 -0.4013450 -0.18302312
#>  [6,] -0.2704635 -0.3098907 -0.4264304 -0.2547432 -0.3817381 -0.11189871
#>  [7,] -0.2725659 -0.4440380 -0.3429243 -0.1178362 -0.3771610 -0.09921673
#>  [8,] -0.3381062 -0.4180137 -0.2800921 -0.2345332 -0.3811412 -0.18476470
#>  [9,] -0.3327223 -0.3443755 -0.2294304 -0.2450957 -0.2653012 -0.24761707
#> [10,] -0.3582532 -0.3323423 -0.3778169 -0.2281920 -0.3075945 -0.17471926
#>               V8          V9
#>  [1,] -0.1335436 -0.15789737
#>  [2,] -0.1239404 -0.03445615
#>  [3,] -0.1577464 -0.20434365
#>  [4,] -0.1003341 -0.10813356
#>  [5,] -0.2110471 -0.05418392
#>  [6,] -0.1035642 -0.05822985
#>  [7,] -0.2587737 -0.25428777
#>  [8,] -0.2048420 -0.10888118
#>  [9,] -0.1172350 -0.18552191
#> [10,] -0.1638307 -0.05774351
#> 
#> $W[[2]]
#>           [,1]        [,2]       [,3]         [,4]       [,5]        [,6]
#> [1,] 0.0216327 0.009479684 0.02746832  0.004488845 0.04524089  0.04306822
#> [2,] 0.1397432 0.044837588 0.10313581 -0.024296080 0.05337269 -0.04874527
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.09946931 -0.06506706 -0.02902391  0.01039985
#> [2,]  0.02865507 -0.00892485 -0.02625266 -0.09597547
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.847852e-05  4.924753e-04  6.533124e-04  7.181596e-04  6.252906e-04
#>  [2,]  4.350395e-05  3.212057e-04  3.904013e-04  4.184360e-04  4.020039e-04
#>  [3,]  5.413979e-05  4.809535e-04  5.911548e-04  6.141650e-04  5.867337e-04
#>  [4,] -2.252798e-05 -4.721579e-04 -6.834483e-04 -7.276726e-04 -5.531347e-04
#>  [5,]  2.349443e-05  9.014937e-05  7.425888e-05  8.365019e-05  1.071918e-04
#>  [6,] -3.724643e-05 -7.548360e-04 -1.051686e-03 -1.093262e-03 -9.036913e-04
#>  [7,]  7.208979e-05  8.431712e-04  1.084716e-03  1.143685e-03  1.051238e-03
#>  [8,]  6.236723e-05  5.119952e-04  6.420762e-04  6.685071e-04  6.282646e-04
#>  [9,]  1.316671e-05 -2.035453e-05 -7.181140e-05 -7.186064e-05 -2.275448e-05
#> [10,] -1.913362e-05 -5.081666e-04 -7.128008e-04 -7.313653e-04 -6.002440e-04
#>                 V14          V15          V16          V17          V18
#>  [1,]  3.797464e-04 1.956007e-04 0.0002654001 0.0003566134 0.0004547123
#>  [2,]  3.274821e-04 3.370722e-04 0.0004125788 0.0005213092 0.0006343238
#>  [3,]  4.497645e-04 3.919814e-04 0.0004448853 0.0005317679 0.0006359299
#>  [4,] -1.574251e-04 2.237664e-04 0.0003691349 0.0004079172 0.0003764914
#>  [5,]  1.620962e-04 2.766140e-04 0.0003761452 0.0004750998 0.0005570803
#>  [6,] -4.019407e-04 8.340948e-05 0.0002067728 0.0002441373 0.0002074198
#>  [7,]  7.145429e-04 4.731599e-04 0.0005463888 0.0006921768 0.0008632842
#>  [8,]  4.619895e-04 3.926364e-04 0.0004401163 0.0005451436 0.0006731475
#>  [9,]  9.356991e-05 2.541167e-04 0.0003582562 0.0004417342 0.0004999303
#> [10,] -2.406497e-04 1.360279e-04 0.0002126465 0.0002675900 0.0002910216
#>                 V19            V2           V20           V21           V22
#>  [1,]  6.252375e-04  4.912603e-05  0.0008042518  0.0007879668  7.027656e-04
#>  [2,]  7.017814e-04  3.852097e-05  0.0008028350  0.0007830635  7.495897e-04
#>  [3,]  7.733088e-04  5.915398e-05  0.0009779474  0.0009463388  8.358654e-04
#>  [4,]  6.712334e-05 -4.204571e-05 -0.0002466612 -0.0001801808  1.345730e-05
#>  [5,]  5.117060e-04  1.326898e-05  0.0004744685  0.0004867916  5.481447e-04
#>  [6,] -2.230853e-04 -7.769948e-05 -0.0006916288 -0.0006160752 -2.801943e-04
#>  [7,]  1.156329e-03  9.535286e-05  0.0015414599  0.0014909861  1.285065e-03
#>  [8,]  8.377560e-04  6.206578e-05  0.0010726082  0.0010108109  8.654353e-04
#>  [9,]  4.011398e-04  1.612881e-06  0.0003114899  0.0003323228  4.211399e-04
#> [10,]  1.542501e-05 -5.650343e-05 -0.0003306512 -0.0002760496 -2.580034e-05
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0004812520 0.0002276449 0.0001711580 0.0001526008 0.0002028128
#>  [2,] 0.0007636603 0.0008186069 0.0008858471 0.0008717814 0.0008790018
#>  [3,] 0.0007501736 0.0006856336 0.0007224918 0.0007334915 0.0007699387
#>  [4,] 0.0003910987 0.0008088622 0.0009687476 0.0010153060 0.0009431242
#>  [5,] 0.0006939885 0.0008826691 0.0009810258 0.0009787697 0.0009663713
#>  [6,] 0.0003079382 0.0009327228 0.0011270827 0.0011391976 0.0010574042
#>  [7,] 0.0010045526 0.0007178662 0.0007044803 0.0007060499 0.0007646194
#>  [8,] 0.0007673752 0.0006978960 0.0007328500 0.0007195804 0.0007564497
#>  [9,] 0.0006112124 0.0008409885 0.0009520015 0.0009578117 0.0009342774
#> [10,] 0.0004412907 0.0009760704 0.0011568594 0.0011478034 0.0010460354
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0003320499 0.0002069113  5.328345e-05 0.0002798199 2.471534e-05
#>  [2,] 0.0009219671 0.0008681809  3.931365e-05 0.0007949570 7.500428e-04
#>  [3,] 0.0008776404 0.0007656610  6.421504e-05 0.0007519431 6.238459e-04
#>  [4,] 0.0007235522 0.0009121182 -4.650778e-05 0.0007405621 1.043692e-03
#>  [5,] 0.0009313538 0.0009599979  1.212359e-05 0.0008368246 9.072891e-04
#>  [6,] 0.0007832879 0.0011002463 -8.981625e-05 0.0008210968 1.282660e-03
#>  [7,] 0.0009599955 0.0007335303  1.074783e-04 0.0008181958 4.896198e-04
#>  [8,] 0.0008858761 0.0007720970  6.467525e-05 0.0007715077 6.347796e-04
#>  [9,] 0.0008662852 0.0009267088 -1.125144e-08 0.0007978687 9.065634e-04
#> [10,] 0.0008194788 0.0010525868 -6.578215e-05 0.0007888872 1.126680e-03
#>                V32          V33           V34           V35           V36
#>  [1,] 0.0001370659 0.0001050530 -0.0001460089 -0.0004400477 -0.0006489084
#>  [2,] 0.0006634092 0.0006400156  0.0006089850  0.0005037094  0.0003896630
#>  [3,] 0.0006198570 0.0005863842  0.0004779426  0.0002871649  0.0000984307
#>  [4,] 0.0006702606 0.0007701629  0.0011103516  0.0015013023  0.0017709076
#>  [5,] 0.0007156902 0.0007198702  0.0008007684  0.0008428414  0.0008551552
#>  [6,] 0.0007884765 0.0008966359  0.0013544139  0.0019028762  0.0023218770
#>  [7,] 0.0006189887 0.0005485833  0.0002527378 -0.0001821542 -0.0005505753
#>  [8,] 0.0006391502 0.0006053910  0.0005104569  0.0002881233  0.0000514032
#>  [9,] 0.0006842362 0.0007046024  0.0008255636  0.0009312792  0.0009957320
#> [10,] 0.0007302006 0.0008182189  0.0011669860  0.0015635650  0.0018363854
#>                 V37          V38          V39            V4           V40
#>  [1,] -4.529041e-04 1.568778e-05 0.0001397000  1.510193e-04 -3.761245e-05
#>  [2,]  3.340009e-04 3.373099e-04 0.0003242613  5.941723e-05  3.302626e-04
#>  [3,]  1.144327e-04 3.082865e-04 0.0003354606  1.217671e-04  2.532563e-04
#>  [4,]  1.434646e-03 5.234926e-04 0.0002426781 -1.833176e-04  5.396253e-04
#>  [5,]  7.047973e-04 4.073467e-04 0.0003143598 -2.046885e-05  4.300675e-04
#>  [6,]  1.825774e-03 5.835616e-04 0.0002275781 -2.951133e-04  6.810637e-04
#>  [7,] -3.839120e-04 2.467725e-04 0.0003918516  2.460786e-04  1.484943e-04
#>  [8,]  4.844242e-05 2.974661e-04 0.0003352531  1.314741e-04  2.446348e-04
#>  [9,]  8.114559e-04 4.069497e-04 0.0002845012 -5.114316e-05  4.329306e-04
#> [10,]  1.423357e-03 4.830419e-04 0.0002291101 -2.196215e-04  6.052020e-04
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0001895414 2.176561e-04  2.304647e-04  2.421027e-04  4.128177e-04
#>  [2,] 0.0003937302 3.818110e-04  2.997854e-04  2.395229e-04  2.550289e-04
#>  [3,] 0.0003947940 4.098093e-04  3.489133e-04  3.015351e-04  4.128477e-04
#>  [4,] 0.0002292538 7.560173e-05 -8.344621e-05 -1.888268e-04 -4.981785e-04
#>  [5,] 0.0003778207 3.155295e-04  2.032065e-04  1.298943e-04  2.924582e-05
#>  [6,] 0.0002427721 6.812357e-05 -1.264293e-04 -2.691691e-04 -7.457805e-04
#>  [7,] 0.0004815446 5.276474e-04  4.918356e-04  4.654255e-04  7.485642e-04
#>  [8,] 0.0004112568 4.380528e-04  3.733673e-04  3.142185e-04  4.456928e-04
#>  [9,] 0.0003405632 2.648881e-04  1.459587e-04  6.613494e-05 -7.139538e-05
#> [10,] 0.0002860182 1.762077e-04  1.569726e-05 -1.177350e-04 -4.997375e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  3.017483e-04  2.068173e-04  1.607990e-04  1.112800e-04  1.239685e-04
#>  [2,]  2.361285e-04  1.793071e-04  1.141291e-04  6.752044e-05  7.291043e-05
#>  [3,]  3.437398e-04  2.358774e-04  1.636166e-04  1.029498e-04  1.132313e-04
#>  [4,] -3.414327e-04 -1.685742e-04 -1.400864e-04 -1.118658e-04 -1.268579e-04
#>  [5,]  6.867082e-05  8.653247e-05  4.450092e-05  1.699661e-05  1.196658e-05
#>  [6,] -5.136686e-04 -2.660655e-04 -2.314479e-04 -1.759329e-04 -2.065403e-04
#>  [7,]  5.839629e-04  3.799598e-04  2.819407e-04  1.881232e-04  2.087868e-04
#>  [8,]  3.811699e-04  2.526876e-04  1.675036e-04  1.088958e-04  1.221482e-04
#>  [9,] -1.146638e-05  3.469801e-05  7.351046e-06 -8.143252e-06 -1.614707e-05
#> [10,] -3.330306e-04 -1.628867e-04 -1.544825e-04 -1.188540e-04 -1.357412e-04
#>                 V50           V51           V52          V53           V54
#>  [1,]  2.446255e-05  4.011552e-05  3.123829e-05 2.125706e-06  7.063764e-06
#>  [2,]  3.058911e-05  1.597363e-05  1.791940e-05 1.691369e-05  1.470571e-05
#>  [3,]  3.791323e-05  3.153769e-05  2.994971e-05 1.695221e-05  1.738315e-05
#>  [4,] -9.938738e-06 -4.469892e-05 -3.297620e-05 1.168351e-05  2.144963e-06
#>  [5,]  1.859756e-05 -3.325891e-06  1.709385e-06 1.544112e-05  1.075258e-05
#>  [6,] -2.398000e-05 -7.168837e-05 -5.311117e-05 1.224392e-05 -1.831884e-06
#>  [7,]  5.637904e-05  6.309222e-05  5.489322e-05 1.676003e-05  2.221126e-05
#>  [8,]  4.189419e-05  3.249699e-05  3.287543e-05 1.864353e-05  1.880377e-05
#>  [9,]  1.207190e-05 -1.116919e-05 -4.920437e-06 1.377468e-05  8.432181e-06
#> [10,] -1.081755e-05 -5.393950e-05 -3.730512e-05 1.387694e-05  1.810730e-06
#>                 V55           V56           V57           V58           V59
#>  [1,]  6.831190e-06  7.772829e-06 -2.921361e-06  1.314309e-05  1.009425e-05
#>  [2,]  1.270947e-05  6.244356e-06  7.548595e-06  7.873757e-06  8.391659e-06
#>  [3,]  1.532705e-05  9.089270e-06  6.546952e-06  1.345163e-05  1.268776e-05
#>  [4,]  2.616248e-06 -1.481264e-06  1.222516e-05 -9.374110e-06 -5.539590e-06
#>  [5,]  9.064873e-06  3.379269e-06  9.349180e-06  1.990706e-06  3.650574e-06
#>  [6,] -2.881183e-06 -7.574118e-06  1.387607e-05 -1.844999e-05 -1.231132e-05
#>  [7,]  2.056007e-05  1.454422e-05  3.136211e-06  2.323777e-05  2.057556e-05
#>  [8,]  1.667899e-05  8.328919e-06  6.596220e-06  1.355516e-05  1.351939e-05
#>  [9,]  7.195846e-06  2.100597e-06  9.345513e-06 -4.177769e-07  1.621793e-06
#> [10,] -8.937747e-08 -5.712614e-06  1.145727e-05 -1.344292e-05 -8.875663e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.189265e-04  7.567850e-06  1.148388e-04  2.550961e-04  4.313717e-04
#>  [2,]  1.265339e-04  6.536994e-06  1.424051e-04  1.841450e-04  2.961695e-04
#>  [3,]  1.579433e-04  9.592093e-06  1.699162e-04  2.572030e-04  4.389097e-04
#>  [4,] -2.432807e-05 -2.835673e-06  2.167058e-05 -1.652665e-04 -4.218307e-04
#>  [5,]  7.984484e-05  3.184926e-06  1.028240e-04  8.140969e-05  8.573301e-05
#>  [6,] -9.159134e-05 -7.801000e-06 -4.162606e-05 -3.003042e-04 -6.655711e-04
#>  [7,]  2.412838e-04  1.496790e-05  2.393304e-04  4.321029e-04  7.677668e-04
#>  [8,]  1.614395e-04  9.971103e-06  1.695493e-04  2.667852e-04  4.793034e-04
#>  [9,]  5.733450e-05  1.813572e-06  8.319811e-05  3.272341e-05 -1.292251e-05
#> [10,] -4.092213e-05 -5.803223e-06 -1.341199e-06 -1.985379e-04 -4.477683e-04
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.0235471  0.02325103  0.02575524  0.02834911  0.02293632  0.02291951
#> [2,] -0.0235471 -0.02325103 -0.02575524 -0.02834911 -0.02293632 -0.02291951
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.02579433  0.02021083  0.02336163  0.02019221
#> [2,] -0.02579433 -0.02021083 -0.02336163 -0.02019221
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.2037807 0.5238539 0.7032939 0.2854188 0.3700090 0.4478230 0.3913455
#>  [8] 0.3180117 0.4585825 0.7070692
#> 
#> $B[[2]]
#> [1] 0.07336246 0.04439480
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0001879424 0.0012476418 0.0010994239 0.0015110079 0.0013841261
#>  [6] 0.0016723236 0.0010099910 0.0011140338 0.0013573163 0.0016039454
#> 
#> $vB[[2]]
#> [1]  0.02252177 -0.02252177
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#>  [2,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#>  [3,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#>  [4,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [5,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [6,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#>  [7,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#>  [8,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#>  [9,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [10,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [11,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [12,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [13,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [14,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [15,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [16,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [17,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [18,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [19,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [20,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [21,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [22,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [23,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [24,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [25,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [26,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [27,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [28,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [29,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [30,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [31,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [32,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [33,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [34,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [35,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [36,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [37,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#> [38,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [39,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [40,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#>  [2,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#>  [3,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#>  [4,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [5,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [6,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#>  [7,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#>  [8,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#>  [9,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [10,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [11,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [12,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [13,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [14,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [15,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [16,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [17,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [18,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [19,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [20,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [21,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [22,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [23,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [24,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [25,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [26,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [27,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [28,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [29,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [30,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [31,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [32,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [33,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [34,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [35,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [36,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [37,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#> [38,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [39,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [40,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#>  [2,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#>  [3,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#>  [4,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [5,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [6,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#>  [7,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#>  [8,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#>  [9,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [10,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [11,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [12,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [13,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [14,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [15,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [16,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [17,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [18,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [19,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [20,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [21,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [22,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [23,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [24,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [25,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [26,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [27,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [28,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [29,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [30,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [31,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [32,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [33,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [34,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [35,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [36,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [37,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#> [38,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [39,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [40,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#>  [2,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#>  [3,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#>  [4,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [5,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [6,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#>  [7,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#>  [8,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#>  [9,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [10,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [11,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [12,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [13,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [14,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [15,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [16,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [17,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [18,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [19,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [20,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [21,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [22,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [23,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [24,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [25,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [26,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [27,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [28,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [29,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [30,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [31,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [32,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [33,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [34,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [35,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [36,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [37,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#> [38,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [39,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [40,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#>  [2,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#>  [3,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#>  [4,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [5,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [6,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#>  [7,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#>  [8,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#>  [9,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [10,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [11,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [12,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [13,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [14,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [15,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [16,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [17,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [18,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [19,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [20,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [21,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [22,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [23,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [24,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [25,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [26,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [27,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [28,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [29,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [30,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [31,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [32,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [33,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [34,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [35,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [36,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [37,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#> [38,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [39,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [40,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#>  [2,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#>  [3,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#>  [4,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [5,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [6,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#>  [7,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#>  [8,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#>  [9,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [10,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [11,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [12,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [13,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [14,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [15,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [16,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [17,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [18,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [19,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [20,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [21,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [22,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [23,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [24,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [25,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [26,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [27,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [28,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [29,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [30,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [31,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [32,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [33,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [34,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [35,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [36,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [37,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#> [38,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [39,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [40,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4452702 0.7589719 0.7949834 0.6465625 0.7222233 0.6463715 0.6874528
#>  [2,] 0.4123807 0.7683652 0.8176328 0.5865546 0.7200584 0.6142926 0.6798129
#>  [3,] 0.3583343 0.6864950 0.7716778 0.4453428 0.6077007 0.5608349 0.6057918
#>  [4,] 0.2601558 0.6450115 0.6571014 0.3039007 0.3729640 0.4172059 0.4211292
#>  [5,] 0.2892709 0.6897077 0.7519795 0.4524109 0.5104577 0.5007119 0.5376319
#>  [6,] 0.4036745 0.7669106 0.8142187 0.6110053 0.7266444 0.6132254 0.6913096
#>  [7,] 0.3748905 0.7486837 0.8122285 0.5592295 0.6683648 0.5968778 0.6574804
#>  [8,] 0.3970519 0.7379614 0.8207060 0.4651032 0.5855123 0.6319367 0.6421137
#>  [9,] 0.3989228 0.7623272 0.8226693 0.5737344 0.6716521 0.6344258 0.6714608
#> [10,] 0.4377617 0.7442236 0.8443875 0.5141661 0.6698082 0.6930635 0.7033583
#> [11,] 0.2998467 0.7448959 0.7753149 0.4695560 0.6213820 0.5432584 0.5997464
#> [12,] 0.3893744 0.7274500 0.8306359 0.4709943 0.6686286 0.6765704 0.6868244
#> [13,] 0.3355255 0.5443004 0.6883660 0.2902630 0.4656937 0.5854283 0.5306938
#> [14,] 0.4727639 0.7902722 0.8483793 0.6159156 0.7600678 0.6828426 0.7332179
#> [15,] 0.4354965 0.8057489 0.8415061 0.6378074 0.7595773 0.6628099 0.7286442
#> [16,] 0.5111346 0.8282587 0.8849135 0.6534820 0.7576798 0.6938232 0.7622008
#> [17,] 0.3890681 0.7442883 0.7817355 0.4835932 0.5859775 0.5725783 0.6108618
#> [18,] 0.2934655 0.7289309 0.7927172 0.4490608 0.5015449 0.5708320 0.5738564
#> [19,] 0.4426970 0.7847617 0.8406069 0.6206365 0.7373256 0.6598323 0.7197106
#> [20,] 0.4524733 0.8061993 0.8593799 0.6107201 0.6966339 0.6462649 0.7181830
#> [21,] 0.3519950 0.7757174 0.8052335 0.5261151 0.6688071 0.5798262 0.6597119
#> [22,] 0.4297661 0.7958820 0.8405761 0.5427202 0.6465042 0.6244162 0.6826059
#> [23,] 0.4709770 0.7911869 0.8453077 0.6120634 0.7652499 0.6881893 0.7326095
#> [24,] 0.4808819 0.6936965 0.8117027 0.4316839 0.6014361 0.6752961 0.6707986
#> [25,] 0.3440874 0.6804106 0.7933166 0.4318932 0.5597565 0.6022757 0.6032471
#> [26,] 0.4196343 0.7166459 0.7995014 0.4969296 0.6400746 0.6378675 0.6348640
#> [27,] 0.4186919 0.8271006 0.8586139 0.6595781 0.7119059 0.6446017 0.7126216
#> [28,] 0.4698484 0.6845628 0.8042137 0.4481196 0.6168002 0.6904651 0.6705314
#> [29,] 0.4716111 0.7429624 0.8195456 0.5207496 0.6735198 0.6349218 0.6734935
#> [30,] 0.5322441 0.7925492 0.8634513 0.6607430 0.7737662 0.7094407 0.7641819
#> [31,] 0.4475132 0.6970112 0.7986928 0.4456466 0.5843442 0.6813996 0.6790336
#> [32,] 0.3769761 0.7167518 0.7601164 0.5152436 0.5732501 0.5418707 0.6113063
#> [33,] 0.4540536 0.7601983 0.7950570 0.6510476 0.7315254 0.6483708 0.6929382
#> [34,] 0.3029338 0.7448216 0.7753952 0.5190897 0.5256878 0.4872705 0.5578490
#> [35,] 0.3129750 0.6003748 0.6842890 0.3391860 0.5205475 0.5650574 0.5380312
#> [36,] 0.3794795 0.6839882 0.7949440 0.4114261 0.5506642 0.6593107 0.6269982
#> [37,] 0.3818102 0.6848822 0.8040070 0.4234180 0.6201707 0.6775353 0.6798880
#> [38,] 0.4301988 0.7943456 0.8365901 0.6237635 0.7089204 0.6374349 0.6860653
#> [39,] 0.4550544 0.7028176 0.8101142 0.4996196 0.6477616 0.6938206 0.6820155
#> [40,] 0.4517234 0.8004889 0.8431209 0.6526348 0.7655377 0.6515274 0.7281388
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7820207 0.6561475 0.7121946
#>  [2,] 0.8270813 0.6518794 0.7348298
#>  [3,] 0.6884700 0.5421978 0.7724162
#>  [4,] 0.4171202 0.3685849 0.6968253
#>  [5,] 0.6256208 0.4816577 0.6994190
#>  [6,] 0.8103705 0.6568580 0.7485092
#>  [7,] 0.7626170 0.6054321 0.7405407
#>  [8,] 0.7302010 0.5605721 0.7527508
#>  [9,] 0.7902023 0.6248898 0.7341622
#> [10,] 0.7848190 0.6390328 0.7964876
#> [11,] 0.7539256 0.5494414 0.7423847
#> [12,] 0.7967653 0.6153863 0.7862013
#> [13,] 0.5628766 0.4598852 0.7457541
#> [14,] 0.8575889 0.7164044 0.7768821
#> [15,] 0.8494099 0.6991990 0.7672768
#> [16,] 0.8566340 0.7304561 0.7988491
#> [17,] 0.6687650 0.5546686 0.7347886
#> [18,] 0.6915535 0.5255428 0.7486727
#> [19,] 0.8242331 0.6828528 0.7720177
#> [20,] 0.8011078 0.6686828 0.7790873
#> [21,] 0.7498483 0.5940541 0.7567379
#> [22,] 0.7546903 0.6122103 0.7724633
#> [23,] 0.8630248 0.7180478 0.7815522
#> [24,] 0.7107328 0.6115989 0.7951101
#> [25,] 0.6751678 0.5101894 0.7408486
#> [26,] 0.7585043 0.6113247 0.7462443
#> [27,] 0.8232914 0.6777874 0.7499922
#> [28,] 0.7140681 0.6292250 0.7895435
#> [29,] 0.7724174 0.6503082 0.7765773
#> [30,] 0.8594314 0.7431290 0.7877153
#> [31,] 0.6991015 0.6045472 0.7903482
#> [32,] 0.6177750 0.5503341 0.7444447
#> [33,] 0.7867623 0.6648361 0.7130907
#> [34,] 0.5828008 0.4894075 0.7283797
#> [35,] 0.6004518 0.4693878 0.7398781
#> [36,] 0.6854616 0.5447089 0.7680697
#> [37,] 0.7519951 0.5954227 0.7882971
#> [38,] 0.8138610 0.6593716 0.7485045
#> [39,] 0.7418746 0.6279611 0.7699408
#> [40,] 0.8579355 0.7098163 0.7498483
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5674851 0.4325149
#>  [2,] 0.5663254 0.4336746
#>  [3,] 0.5635986 0.4364014
#>  [4,] 0.5557698 0.4442302
#>  [5,] 0.5603752 0.4396248
#>  [6,] 0.5674267 0.4325733
#>  [7,] 0.5656890 0.4343110
#>  [8,] 0.5640325 0.4359675
#>  [9,] 0.5664605 0.4335395
#> [10,] 0.5681389 0.4318611
#> [11,] 0.5636547 0.4363453
#> [12,] 0.5671829 0.4328171
#> [13,] 0.5604006 0.4395994
#> [14,] 0.5695423 0.4304577
#> [15,] 0.5696387 0.4303613
#> [16,] 0.5702989 0.4297011
#> [17,] 0.5629178 0.4370822
#> [18,] 0.5637063 0.4362937
#> [19,] 0.5690542 0.4309458
#> [20,] 0.5680080 0.4319920
#> [21,] 0.5654044 0.4345956
#> [22,] 0.5658568 0.4341432
#> [23,] 0.5699180 0.4300820
#> [24,] 0.5650290 0.4349710
#> [25,] 0.5629912 0.4370088
#> [26,] 0.5654220 0.4345780
#> [27,] 0.5688376 0.4311624
#> [28,] 0.5662234 0.4337766
#> [29,] 0.5659136 0.4340864
#> [30,] 0.5704769 0.4295231
#> [31,] 0.5656810 0.4343190
#> [32,] 0.5631802 0.4368198
#> [33,] 0.5675874 0.4324126
#> [34,] 0.5621530 0.4378470
#> [35,] 0.5614291 0.4385709
#> [36,] 0.5646580 0.4353420
#> [37,] 0.5661044 0.4338956
#> [38,] 0.5679402 0.4320598
#> [39,] 0.5669345 0.4330655
#> [40,] 0.5687812 0.4312188
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]      [,3]         [,4]        [,5]         [,6]
#>  [1,] -0.21979984 1.1470514 1.3552304  0.603963319  0.95551701  0.603127409
#>  [2,] -0.35413254 1.1991031 1.5003909  0.349740416  0.94475131  0.465392101
#>  [3,] -0.58260099 0.7837834 1.2178091 -0.219505898  0.43765738  0.244550977
#>  [4,] -1.04515891 0.5971826 0.6504035 -0.828791084 -0.51952224 -0.334254093
#>  [5,] -0.89892790 0.7987529 1.1091976 -0.190934462  0.04183705  0.002847685
#>  [6,] -0.39017772 1.1909484 1.4776588  0.451539935  0.97766369  0.460890504
#>  [7,] -0.51129306 1.0916044 1.4645562  0.238035567  0.70079856  0.392472779
#>  [8,] -0.41776421 1.0353996 1.5211379 -0.139814380  0.34544394  0.540534186
#>  [9,] -0.40995540 1.1654810 1.5345382  0.297104125  0.71566674  0.551251270
#> [10,] -0.25025119 1.0680379 1.6912427  0.056679431  0.70731774  0.814480673
#> [11,] -0.84802791 1.0715730 1.2385695 -0.121926843  0.49541841  0.173467446
#> [12,] -0.44994262 0.9817232 1.5901410 -0.116153093  0.70198905  0.738055120
#> [13,] -0.68329845 0.1776675 0.7924911 -0.894106831 -0.13744109  0.345097979
#> [14,] -0.10905242 1.3265670 1.7219456  0.472247659  1.15305138  0.766865984
#> [15,] -0.25945987 1.4226203 1.6694770  0.565860075  1.15036353  0.675841387
#> [16,]  0.04454585 1.5733361 2.0398059  0.634380389  1.14000070  0.818054584
#> [17,] -0.45123107 1.0683779 1.2758091 -0.065650806  0.34736114  0.292378344
#> [18,] -0.87861203 0.9892050 1.3413825 -0.204466025  0.00617963  0.285246600
#> [19,] -0.23022333 1.2936344 1.6627507  0.492250608  1.03211451  0.662547024
#> [20,] -0.19068264 1.4255008 1.8101489  0.450340262  0.83131978  0.602661161
#> [21,] -0.61028140 1.2408812 1.4193311  0.104555500  0.70279467  0.322059985
#> [22,] -0.28280533 1.3607527 1.6625210  0.171298371  0.60370817  0.508335647
#> [23,] -0.11622280 1.3320945 1.6982624  0.455994007  1.18168111  0.791667498
#> [24,] -0.07650964 0.8174581 1.4611122 -0.274983980  0.41145234  0.732237471
#> [25,] -0.64513176 0.7556596 1.3450341 -0.274131109  0.24017370  0.414956153
#> [26,] -0.32427478 0.9278846 1.3831812 -0.012281812  0.57568807  0.566120605
#> [27,] -0.32814565 1.5652164 1.8038249  0.661414723  0.90465867  0.595393353
#> [28,] -0.12075302 0.7748207 1.4128410 -0.208271302  0.47598848  0.802294770
#> [29,] -0.11367772 1.0614229 1.5132716  0.083045997  0.72414800  0.553390206
#> [30,]  0.12915550 1.3403603 1.8442561  0.666606822  1.22970105  0.892669278
#> [31,] -0.21072369 0.8331055 1.3781443 -0.218276219  0.34063283  0.760211427
#> [32,] -0.50240298 0.9284060 1.1533177  0.060993250  0.29512412  0.167875857
#> [33,] -0.18430556 1.1537672 1.3556817  0.623647308  1.00237561  0.611885329
#> [34,] -0.83336631 1.0711820 1.2390306  0.076396069  0.10284169 -0.050929024
#> [35,] -0.78624732 0.4070272 0.7735532 -0.666923591  0.08223632  0.261713418
#> [36,] -0.49175823 0.7721611 1.3549887 -0.358072944  0.20335471  0.660223879
#> [37,] -0.48187184 0.7763004 1.4115293 -0.308757752  0.49027273  0.742467872
#> [38,] -0.28104023 1.3513217 1.6330721  0.505553185  0.89014663  0.564248022
#> [39,] -0.18026889 0.8607513 1.4507527 -0.001521716  0.60921437  0.818042087
#> [40,] -0.19371002 1.3893529 1.6816349  0.630640955  1.18328383  0.625759985
#>             [,7]       [,8]        [,9]     [,10]
#>  [1,]  0.7882377  1.2774811  0.64617291 0.9060669
#>  [2,]  0.7529122  1.5650816  0.62731047 1.0192669
#>  [3,]  0.4296571  0.7929759  0.16919358 1.2220051
#>  [4,] -0.3181396 -0.3346065 -0.53829249 0.8322256
#>  [5,]  0.1508129  0.5134753 -0.07340209 0.8445328
#>  [6,]  0.8062490  1.4524196  0.64932368 1.0906771
#>  [7,]  0.6520860  1.1670810  0.42815108 1.0487806
#>  [8,]  0.5845503  0.9956424  0.24348413 1.1133376
#>  [9,]  0.7147994  1.3261454  0.51035534 1.0158435
#> [10,]  0.8633414  1.2939736  0.57116866 1.3644847
#> [11,]  0.4044087  1.1196596  0.19841396 1.0584002
#> [12,]  0.7853144  1.3661986  0.47001087 1.3021779
#> [13,]  0.1229298  0.2528447 -0.16080480 1.0760941
#> [14,]  1.0110106  1.7954067  0.92669533 1.2475882
#> [15,]  0.9877549  1.7299805  0.84348635 1.1929981
#> [16,]  1.1647835  1.7876101  0.99693780 1.3791169
#> [17,]  0.4509363  0.7026047  0.21955199 1.0190556
#> [18,]  0.2976029  0.8073922  0.10226039 1.0915461
#> [19,]  0.9430264  1.5452948  0.76691345 1.2197393
#> [20,]  0.9354665  1.3932328  0.70223342 1.2603554
#> [21,]  0.6620108  1.0978036  0.38075061 1.1348775
#> [22,]  0.7657735  1.1237859  0.45661250 1.2222730
#> [23,]  1.0079027  1.8406438  0.93479859 1.2747351
#> [24,]  0.7117990  0.8989458  0.45403821 1.3560076
#> [25,]  0.4190133  0.7316525  0.04076323 1.0503839
#> [26,]  0.5531410  1.1444967  0.45288376 1.0786809
#> [27,]  0.9081509  1.5388078  0.74362211 1.0985706
#> [28,]  0.7105895  0.9152249  0.52889370 1.3221758
#> [29,]  0.7240286  1.2220115  0.62039434 1.2458304
#> [30,]  1.1757453  1.8105757  1.06229559 1.3112084
#> [31,]  0.7493342  0.8430228  0.42444829 1.3270253
#> [32,]  0.4528067  0.4801150  0.20202087 1.0691998
#> [33,]  0.8138918  1.3055188  0.68492077 0.9104425
#> [34,]  0.2324370  0.3342818 -0.04237653 0.9864175
#> [35,]  0.1524193  0.4073481 -0.12260199 1.0453351
#> [36,]  0.5193605  0.7789865  0.17931466 1.1974436
#> [37,]  0.7532569  1.1092815  0.38642856 1.3146909
#> [38,]  0.7817876  1.4752956  0.66049530 1.0906523
#> [39,]  0.7630497  1.0557346  0.52347977 1.2079768
#> [40,]  0.9852001  1.7982473  0.89449220 1.0978032
#> 
#> $pre[[3]]
#>            [,1]        [,2]
#>  [1,] 0.2286242 -0.04297345
#>  [2,] 0.2236745 -0.04319997
#>  [3,] 0.2137242 -0.04205565
#>  [4,] 0.1961537 -0.02785764
#>  [5,] 0.2052461 -0.03743885
#>  [6,] 0.2245262 -0.04683381
#>  [7,] 0.2204527 -0.04383068
#>  [8,] 0.2183419 -0.03920222
#>  [9,] 0.2231462 -0.04427834
#> [10,] 0.2269699 -0.04729184
#> [11,] 0.2103044 -0.04570333
#> [12,] 0.2214179 -0.04894880
#> [13,] 0.2012711 -0.04151704
#> [14,] 0.2325894 -0.04739476
#> [15,] 0.2309174 -0.04945997
#> [16,] 0.2378062 -0.04526479
#> [17,] 0.2173783 -0.03563412
#> [18,] 0.2084290 -0.04778890
#> [19,] 0.2302556 -0.04773756
#> [20,] 0.2298494 -0.04387915
#> [21,] 0.2187167 -0.04440866
#> [22,] 0.2252129 -0.03975354
#> [23,] 0.2328893 -0.04862715
#> [24,] 0.2228175 -0.03878018
#> [25,] 0.2130570 -0.04025354
#> [26,] 0.2214704 -0.04172662
#> [27,] 0.2299754 -0.04713467
#> [28,] 0.2238096 -0.04264941
#> [29,] 0.2258520 -0.03934567
#> [30,] 0.2381307 -0.04566662
#> [31,] 0.2202313 -0.04401986
#> [32,] 0.2154997 -0.03857923
#> [33,] 0.2293399 -0.04267459
#> [34,] 0.2125279 -0.03737650
#> [35,] 0.2034130 -0.04355113
#> [36,] 0.2157570 -0.04433145
#> [37,] 0.2161596 -0.04981496
#> [38,] 0.2288367 -0.04461540
#> [39,] 0.2253903 -0.04396450
#> [40,] 0.2307746 -0.04610566
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5674851  0.5674851
#>  [2,]  0.4336746 -0.4336746
#>  [3,] -0.5635986  0.5635986
#>  [4,]  0.4442302 -0.4442302
#>  [5,]  0.4396248 -0.4396248
#>  [6,]  0.4325733 -0.4325733
#>  [7,]  0.4343110 -0.4343110
#>  [8,] -0.5640325  0.5640325
#>  [9,]  0.4335395 -0.4335395
#> [10,]  0.4318611 -0.4318611
#> [11,]  0.4363453 -0.4363453
#> [12,]  0.4328171 -0.4328171
#> [13,]  0.4395994 -0.4395994
#> [14,]  0.4304577 -0.4304577
#> [15,]  0.4303613 -0.4303613
#> [16,] -0.5702989  0.5702989
#> [17,] -0.5629178  0.5629178
#> [18,]  0.4362937 -0.4362937
#> [19,] -0.5690542  0.5690542
#> [20,] -0.5680080  0.5680080
#> [21,]  0.4345956 -0.4345956
#> [22,]  0.4341432 -0.4341432
#> [23,]  0.4300820 -0.4300820
#> [24,]  0.4349710 -0.4349710
#> [25,] -0.5629912  0.5629912
#> [26,] -0.5654220  0.5654220
#> [27,] -0.5688376  0.5688376
#> [28,]  0.4337766 -0.4337766
#> [29,] -0.5659136  0.5659136
#> [30,] -0.5704769  0.5704769
#> [31,] -0.5656810  0.5656810
#> [32,] -0.5631802  0.5631802
#> [33,] -0.5675874  0.5675874
#> [34,] -0.5621530  0.5621530
#> [35,]  0.4385709 -0.4385709
#> [36,]  0.4353420 -0.4353420
#> [37,] -0.5661044  0.5661044
#> [38,] -0.5679402  0.5679402
#> [39,] -0.5669345  0.5669345
#> [40,]  0.4312188 -0.4312188
#> 
#> $L
#> [1] 0.6951235 0.7181409 0.8373725 0.7456541 0.6939691 0.6962170
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5507246