Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11         V12          V13
#>  [1,] -0.3368083 -0.050679219 -0.05994036 -0.03795938 -0.060097116
#>  [2,] -0.3256976  0.006666339 -0.10411532 -0.05871344 -0.163105570
#>  [3,] -0.3378858 -0.200319116 -0.10074588 -0.09262171 -0.157934229
#>  [4,] -0.3873798 -0.196171064 -0.22927903 -0.03822660 -0.171609832
#>  [5,] -0.1832429 -0.134683403 -0.09200227 -0.06461092 -0.016568214
#>  [6,] -0.3174836 -0.031562994 -0.03270095 -0.16942673 -0.193380504
#>  [7,] -0.3484506 -0.207528222 -0.04480106 -0.17878708 -0.053427106
#>  [8,] -0.3434726 -0.093851076 -0.13240135 -0.08815129 -0.000393822
#>  [9,] -0.2698943 -0.084332899 -0.07613925 -0.10983891 -0.007619615
#> [10,] -0.2249669 -0.186501489 -0.08215628 -0.01892805 -0.025592840
#>                 V14         V15          V16           V17          V18
#>  [1,] -0.1797483040 -0.09727466 -0.173186550 -0.1030288162 -0.162958610
#>  [2,]  0.0159142119 -0.02860833 -0.019378300  0.0048454869  0.032094595
#>  [3,] -0.1646402568  0.02001100 -0.065818836  0.0009534149  0.002226094
#>  [4,] -0.0957004793 -0.12582274 -0.066501121 -0.0387010540 -0.031896892
#>  [5,]  0.0223862648 -0.03813829  0.002641806  0.0061361918 -0.055754020
#>  [6,] -0.1592482103 -0.11655290 -0.132606400  0.0252682818 -0.054939353
#>  [7,]  0.0607458354 -0.07180160 -0.026463143  0.0647045737  0.053341474
#>  [8,] -0.0717974618 -0.09359922 -0.016043501 -0.1160540941  0.112654246
#>  [9,]  0.0004482816 -0.12593765 -0.067002881  0.0379969874 -0.084980042
#> [10,] -0.0668592976 -0.11406110 -0.026052340 -0.0399337155  0.067332319
#>               V19         V2         V20          V21          V22          V23
#>  [1,] -0.15579753 -0.2938148 -0.09202490 -0.111335882 -0.038744892 -0.008832731
#>  [2,]  0.01084744 -0.3055323 -0.03130111  0.004080555  0.003383667  0.025011106
#>  [3,] -0.09216363 -0.2678869 -0.04744734  0.037635113  0.002654509  0.131947194
#>  [4,]  0.08189700 -0.3660765  0.02799081  0.082724081  0.091759071  0.118075796
#>  [5,]  0.06389167 -0.2923243 -0.02095771  0.076152745  0.008101687  0.032529040
#>  [6,] -0.09501145 -0.2610601  0.01805961  0.012723428 -0.096461634 -0.068262042
#>  [7,]  0.04915090 -0.3772510  0.01694145  0.084158704  0.184354795  0.170212277
#>  [8,]  0.02142051 -0.1560850  0.07340932  0.079965000  0.053279125  0.215357715
#>  [9,]  0.08607962 -0.2281247  0.06906256  0.119811452  0.200249453  0.045165690
#> [10,] -0.09900096 -0.3261856 -0.02036910 -0.008545712  0.062887518 -0.033055117
#>                V24         V25        V26         V27          V28         V29
#>  [1,] -0.039028074  0.06396233 0.06997731 -0.08718922 -0.017397649  0.04072768
#>  [2,] -0.009191383 -0.02671699 0.05598758  0.03970560  0.129793990 -0.02562598
#>  [3,]  0.102309804  0.16498680 0.08030469  0.07151746  0.062282127  0.13874446
#>  [4,]  0.117072193  0.03559057 0.19487918  0.13095015  0.144773786  0.18091703
#>  [5,]  0.101302102  0.10638185 0.09622255  0.16980995  0.091893477  0.10262654
#>  [6,]  0.091034616  0.02669270 0.12933233  0.05068733  0.224719415  0.19890791
#>  [7,]  0.147892294 -0.02127909 0.06219592  0.02729003  0.110943652 -0.01036477
#>  [8,]  0.059639030  0.13166025 0.09688158  0.06234040  0.004245189  0.02817191
#>  [9,]  0.158203245  0.24059660 0.07352112  0.17990032  0.081168262  0.16423180
#> [10,]  0.069087188  0.06519505 0.12154664  0.17581530  0.069451170  0.01674787
#>               V3         V30          V31         V32          V33         V34
#>  [1,] -0.3738104 -0.06762477 -0.042500328 -0.08251001 -0.095413724 -0.10935159
#>  [2,] -0.2749512  0.04074341  0.010265734 -0.08581871  0.014795644  0.04404330
#>  [3,] -0.2763070  0.03632579  0.022278248 -0.14223343  0.055387454  0.06826050
#>  [4,] -0.3387543  0.03766118  0.010092591  0.09317551  0.099045146  0.14621483
#>  [5,] -0.2235997  0.12384175  0.150800212  0.05015616  0.052578399 -0.04025306
#>  [6,] -0.2479920  0.03083479  0.013218105  0.06752195 -0.041534109  0.03983260
#>  [7,] -0.3019152  0.08578264  0.007368446 -0.01521754  0.020704541  0.02479194
#>  [8,] -0.2722502  0.16180451 -0.034746653  0.06373572 -0.042572916 -0.03598985
#>  [9,] -0.3251987 -0.03825577  0.009611462  0.02782712 -0.050967500  0.01022481
#> [10,] -0.1946298  0.10620254  0.036794733 -0.01213162  0.004740668 -0.09651746
#>               V35         V36          V37          V38         V39         V4
#>  [1,] -0.08422366 -0.07590001  0.043704675 -0.078973511  0.01042238 -0.2206176
#>  [2,] -0.01042544  0.02681049  0.065398630  0.035030626  0.09526694 -0.2483162
#>  [3,]  0.09841877  0.18978679  0.076892609  0.022456730  0.06353540 -0.2751898
#>  [4,]  0.03377674  0.01171360 -0.021381238 -0.035726299 -0.03843220 -0.2279579
#>  [5,]  0.14273258  0.03482567 -0.057174544 -0.020991426 -0.11681197 -0.2265329
#>  [6,]  0.04469500  0.11523831  0.024213140  0.003193425 -0.01594225 -0.1815214
#>  [7,]  0.08579617  0.07007989  0.067556953 -0.044082780 -0.14039214 -0.4133410
#>  [8,] -0.07888915 -0.00097821  0.025377696  0.022471612 -0.10855570 -0.3258078
#>  [9,]  0.08879832  0.05352570 -0.005427643 -0.019674191 -0.12692928 -0.2380660
#> [10,] -0.03477564 -0.02726116  0.001154660 -0.048909218  0.03709023 -0.2739852
#>               V40          V41          V42         V43          V44
#>  [1,] -0.06056141 -0.100749633 -0.147527562 -0.13456783 -0.213872308
#>  [2,]  0.03762316 -0.048399496 -0.047782814 -0.16973331 -0.119507593
#>  [3,] -0.10846139 -0.031985455 -0.159751245 -0.14506671 -0.188473069
#>  [4,] -0.06946242  0.037002202 -0.093070756 -0.13046732  0.001308259
#>  [5,] -0.06132372 -0.035615216 -0.083572127  0.01584979 -0.063771796
#>  [6,] -0.08006367 -0.056269309  0.037663575 -0.08935043 -0.081305728
#>  [7,] -0.04674557  0.004688796 -0.083615138  0.01817180 -0.172576677
#>  [8,]  0.01555257 -0.032866870 -0.105888754 -0.03966601 -0.099684525
#>  [9,] -0.01046606 -0.149809330  0.003370432 -0.14647437 -0.014504835
#> [10,] -0.11695443 -0.078937147 -0.102391529 -0.06083391 -0.157178615
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.12409981 -0.27958111 -0.1399052 -0.1683941 -0.3290577 -0.2413009
#>  [2,] -0.11006495 -0.21312586 -0.1522696 -0.3187829 -0.2140786 -0.2574219
#>  [3,] -0.21409481 -0.12092692 -0.2733735 -0.1864793 -0.3274594 -0.1962802
#>  [4,] -0.03146964 -0.09833033 -0.1927338 -0.1823561 -0.2731632 -0.2077067
#>  [5,] -0.11339306 -0.13169369 -0.2442203 -0.2004147 -0.2779424 -0.1617558
#>  [6,] -0.19580283 -0.18828047 -0.1521544 -0.2784244 -0.3088237 -0.1516306
#>  [7,] -0.14384925 -0.08798771 -0.2528749 -0.2701051 -0.3479950 -0.2597753
#>  [8,] -0.11581160 -0.06136068 -0.1661080 -0.1837111 -0.1246311 -0.1924608
#>  [9,] -0.14309030 -0.08158394 -0.1665099 -0.2995630 -0.3660705 -0.2141418
#> [10,] -0.15483119 -0.24584416 -0.2385458 -0.1284124 -0.2260438 -0.3458423
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3011705 -0.3975548 -0.2964555 -0.4057229 -0.2218572 -0.2654226
#>  [2,] -0.2427630 -0.3445788 -0.3448489 -0.2194288 -0.4112798 -0.2692969
#>  [3,] -0.3344989 -0.2688506 -0.3839062 -0.3241124 -0.2932912 -0.2696065
#>  [4,] -0.3860760 -0.3022008 -0.2536189 -0.2544358 -0.2969969 -0.3403108
#>  [5,] -0.3872438 -0.2661006 -0.3614819 -0.3715357 -0.3951942 -0.3988353
#>  [6,] -0.2837449 -0.3289774 -0.2905277 -0.3456018 -0.3042530 -0.3217331
#>  [7,] -0.3308251 -0.4156481 -0.4024322 -0.3839331 -0.4218304 -0.2928767
#>  [8,] -0.1978149 -0.3038790 -0.1903775 -0.3493707 -0.2715705 -0.2251324
#>  [9,] -0.2790731 -0.3145131 -0.4299766 -0.3835717 -0.4074517 -0.4588766
#> [10,] -0.3494513 -0.2758783 -0.2861621 -0.2731933 -0.2580406 -0.3944351
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4261154 -0.2952986 -0.3930213 -0.2908781 -0.2272484 -0.3597812
#>  [2,] -0.2655103 -0.2563748 -0.2750820 -0.3377660 -0.1694237 -0.2707528
#>  [3,] -0.2640458 -0.2895706 -0.3343077 -0.2805750 -0.1023098 -0.2267443
#>  [4,] -0.3314720 -0.3582547 -0.3342604 -0.3513089 -0.2636855 -0.3982771
#>  [5,] -0.3037201 -0.3631595 -0.3518147 -0.3731335 -0.2694256 -0.2823225
#>  [6,] -0.3024264 -0.3518079 -0.2191259 -0.3346024 -0.1508195 -0.3243556
#>  [7,] -0.3700560 -0.4823205 -0.3149359 -0.4187088 -0.1628465 -0.4148778
#>  [8,] -0.3534689 -0.2430844 -0.3662881 -0.2356224 -0.1427163 -0.2280269
#>  [9,] -0.3656273 -0.3071659 -0.2879184 -0.3434803 -0.2249753 -0.4203036
#> [10,] -0.3319784 -0.2644789 -0.3969045 -0.3478538 -0.1038347 -0.3003611
#>                V7          V8          V9
#>  [1,] -0.16256172 -0.16915936 -0.06822371
#>  [2,] -0.26334334 -0.10850077 -0.05644752
#>  [3,] -0.18937728 -0.20363208 -0.19939972
#>  [4,] -0.18435720 -0.27220352 -0.21936877
#>  [5,] -0.19287796 -0.09598325 -0.02569129
#>  [6,] -0.25518630 -0.14396963 -0.13519942
#>  [7,] -0.14956387 -0.07310951 -0.14503547
#>  [8,] -0.26181776 -0.12464974 -0.12985117
#>  [9,] -0.15627556 -0.25217438 -0.27559870
#> [10,] -0.09414763 -0.21768791 -0.07875034
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]         [,4]        [,5]       [,6]
#> [1,] 0.004292392 -0.04869204 -0.06889138 -0.005495452  0.04764415 0.05902579
#> [2,] 0.081742004 -0.03676815  0.07155461 -0.040034305 -0.05281560 0.07251191
#>            [,7]        [,8]       [,9]      [,10]
#> [1,] 0.07011492  0.07110305  0.1500667 0.09329442
#> [2,] 0.02664542 -0.03230224 -0.1532701 0.05324956
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.107825e-04  1.018111e-03  1.181071e-03  1.196495e-03  1.204088e-03
#>  [2,]  1.240690e-04  1.071764e-03  1.205625e-03  1.197697e-03  1.264434e-03
#>  [3,]  3.013235e-04  2.577605e-03  2.937042e-03  2.933847e-03  3.080855e-03
#>  [4,]  9.208407e-05  7.601379e-04  8.474053e-04  8.250703e-04  8.929523e-04
#>  [5,]  1.290842e-06  8.940386e-06 -9.747275e-06 -2.666382e-05 -1.673597e-05
#>  [6,]  9.421366e-05  8.102858e-04  8.830965e-04  8.570904e-04  9.268181e-04
#>  [7,]  7.240485e-05  6.208731e-04  6.851345e-04  6.630990e-04  7.090459e-04
#>  [8,] -3.150179e-06 -3.464638e-05 -4.915988e-05 -5.996959e-05 -5.659746e-05
#>  [9,] -3.008215e-04 -2.557130e-03 -2.972673e-03 -2.987860e-03 -3.118503e-03
#> [10,]  4.459326e-05  3.849068e-04  4.297621e-04  4.206258e-04  4.484550e-04
#>                 V14           V15           V16           V17           V18
#>  [1,]  1.154221e-03  1.158445e-03  1.209634e-03  1.324003e-03  1.542459e-03
#>  [2,]  1.292158e-03  1.314362e-03  1.383408e-03  1.500544e-03  1.725458e-03
#>  [3,]  3.086542e-03  3.094418e-03  3.222887e-03  3.486890e-03  4.035249e-03
#>  [4,]  9.425274e-04  9.769125e-04  1.025804e-03  1.107715e-03  1.266581e-03
#>  [5,]  3.877183e-05  7.473873e-05  1.015222e-04  1.200521e-04  1.142802e-04
#>  [6,]  1.000649e-03  1.046825e-03  1.112762e-03  1.207225e-03  1.372433e-03
#>  [7,]  7.496042e-04  7.804651e-04  8.314772e-04  9.040437e-04  1.025198e-03
#>  [8,] -1.232283e-05  1.720348e-05  4.028425e-05  5.544456e-05  3.927471e-05
#>  [9,] -2.956625e-03 -2.868309e-03 -2.922265e-03 -3.133295e-03 -3.681573e-03
#> [10,]  4.847437e-04  5.097355e-04  5.489842e-04  6.040317e-04  6.807772e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  1.971215e-03  9.803357e-05  2.235937e-03  2.383845e-03  2.141880e-03
#>  [2,]  2.166405e-03  1.116322e-04  2.472224e-03  2.627499e-03  2.372806e-03
#>  [3,]  5.137544e-03  2.827104e-04  5.928609e-03  6.308918e-03  5.658551e-03
#>  [4,]  1.576242e-03  8.329032e-05  1.810532e-03  1.916761e-03  1.729240e-03
#>  [5,]  8.480781e-05 -4.719770e-06  6.385905e-05  7.107638e-05  1.270643e-04
#>  [6,]  1.674245e-03  8.022305e-05  1.892753e-03  2.022668e-03  1.866645e-03
#>  [7,]  1.259560e-03  6.345030e-05  1.429065e-03  1.521362e-03  1.397458e-03
#>  [8,] -5.110924e-07 -7.397177e-06 -2.627596e-05 -2.629985e-05  3.942241e-05
#>  [9,] -4.889916e-03 -3.007973e-04 -5.750518e-03 -6.068551e-03 -5.196210e-03
#> [10,]  8.257580e-04  3.691221e-05  9.228619e-04  9.809825e-04  9.290710e-04
#>                 V23           V24           V25           V26           V27
#>  [1,]  1.593947e-03  0.0013780861  0.0013239672  0.0014241352  0.0014741664
#>  [2,]  1.863151e-03  0.0016652366  0.0016001427  0.0017486904  0.0017810034
#>  [3,]  4.462724e-03  0.0039939074  0.0038226728  0.0041761312  0.0042857286
#>  [4,]  1.378441e-03  0.0012421607  0.0012007900  0.0013362975  0.0013444229
#>  [5,]  1.475958e-04  0.0001633287  0.0001750601  0.0001864020  0.0001629759
#>  [6,]  1.507804e-03  0.0013798282  0.0013278190  0.0014363913  0.0014332991
#>  [7,]  1.118660e-03  0.0010182630  0.0009951157  0.0011005863  0.0011106045
#>  [8,]  8.116132e-05  0.0001009145  0.0001170816  0.0001271096  0.0001101157
#>  [9,] -3.863240e-03 -0.0033009483 -0.0031358754 -0.0035230254 -0.0036916952
#> [10,]  7.594618e-04  0.0006930390  0.0006773794  0.0007360707  0.0007364195
#>                 V28           V29            V3           V30           V31
#>  [1,]  1.198278e-03  0.0007332755  1.259455e-04  0.0006748675  3.096004e-05
#>  [2,]  1.505080e-03  0.0010644767  1.492469e-04  0.0010330446  3.576070e-04
#>  [3,]  3.685155e-03  0.0026130230  3.596666e-04  0.0024922211  8.452374e-04
#>  [4,]  1.138747e-03  0.0008390959  1.130705e-04  0.0008387385  3.788690e-04
#>  [5,]  1.220688e-04  0.0001427771  5.300231e-06  0.0001599297  1.710455e-04
#>  [6,]  1.221905e-03  0.0009306849  1.152629e-04  0.0009144997  4.000420e-04
#>  [7,]  9.353191e-04  0.0006888344  8.968028e-05  0.0006871925  3.146848e-04
#>  [8,]  8.060848e-05  0.0001109366 -6.356001e-09  0.0001280824  1.636520e-04
#>  [9,] -3.114585e-03 -0.0019183645 -3.500820e-04 -0.0018405106 -2.338784e-04
#> [10,]  6.091035e-04  0.0004596689  5.564347e-05  0.0004565747  2.196435e-04
#>                 V32           V33           V34           V35           V36
#>  [1,] -2.661155e-05 -2.787476e-04 -6.642860e-04 -7.273780e-04 -7.806232e-04
#>  [2,]  2.558540e-04  6.761559e-05 -2.366686e-04 -2.554233e-04 -3.256456e-04
#>  [3,]  6.449164e-04  1.439134e-04 -5.547199e-04 -6.701998e-04 -9.459171e-04
#>  [4,]  3.184705e-04  2.048497e-04  2.518194e-05  1.417826e-05 -7.872463e-05
#>  [5,]  1.301070e-04  1.609774e-04  1.593482e-04  1.944817e-04  2.368525e-04
#>  [6,]  2.954498e-04  2.172772e-04  8.241281e-06  2.653541e-05  1.800639e-05
#>  [7,]  2.425980e-04  1.455868e-04 -2.280720e-05 -2.162055e-05 -6.290755e-05
#>  [8,]  1.287818e-04  1.508714e-04  1.605499e-04  1.905661e-04  2.224196e-04
#>  [9,] -2.318581e-04  4.350487e-04  1.154435e-03  1.382343e-03  1.868281e-03
#> [10,]  1.610163e-04  9.893244e-05 -1.692298e-05 -7.429960e-06 -1.219442e-05
#>                 V37           V38           V39            V4           V40
#>  [1,] -3.914522e-04  4.928800e-04  7.184705e-04  2.443366e-04  3.161331e-04
#>  [2,] -4.265265e-05  7.796327e-04  9.693358e-04  2.712942e-04  5.466128e-04
#>  [3,] -2.961084e-04  1.805963e-03  2.327243e-03  6.571656e-04  1.230954e-03
#>  [4,]  6.965931e-05  6.158255e-04  7.339945e-04  1.996983e-04  4.233324e-04
#>  [5,]  2.170797e-04  1.061046e-04  5.468431e-05  4.377851e-06  1.054243e-04
#>  [6,]  1.903297e-04  6.966599e-04  7.949672e-04  2.032230e-04  5.362713e-04
#>  [7,]  8.062673e-05  4.949263e-04  5.727911e-04  1.577902e-04  3.503506e-04
#>  [8,]  1.900642e-04  6.531929e-05  1.272042e-05 -4.220233e-06  6.191637e-05
#>  [9,]  1.132634e-03 -1.381575e-03 -2.055133e-03 -6.573663e-04 -7.343393e-04
#> [10,]  7.952548e-05  3.212177e-04  3.646937e-04  1.006696e-04  2.342615e-04
#>                 V41           V42           V43           V44           V45
#>  [1,]  2.497690e-04  2.938665e-04  4.707448e-04  5.579342e-04  7.119694e-04
#>  [2,]  4.475494e-04  4.509062e-04  6.007313e-04  6.516903e-04  7.897031e-04
#>  [3,]  1.013431e-03  1.078779e-03  1.470522e-03  1.603191e-03  1.944317e-03
#>  [4,]  3.492950e-04  3.401824e-04  4.465683e-04  4.721457e-04  5.576559e-04
#>  [5,]  8.985820e-05  4.432359e-05  2.439202e-05  6.109020e-06 -1.616729e-05
#>  [6,]  4.349445e-04  3.912542e-04  4.722681e-04  4.947502e-04  5.867204e-04
#>  [7,]  2.927989e-04  2.735459e-04  3.551308e-04  3.797518e-04  4.517927e-04
#>  [8,]  5.857336e-05  2.511644e-05  9.095707e-06 -1.309728e-05 -4.298262e-05
#>  [9,] -6.280757e-04 -8.477973e-04 -1.350834e-03 -1.563145e-03 -1.990300e-03
#> [10,]  1.982682e-04  1.812520e-04  2.311172e-04  2.378725e-04  2.730503e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  5.880065e-04  4.745126e-04  3.721078e-04  2.316163e-04  3.333311e-04
#>  [2,]  6.380910e-04  4.969460e-04  4.026935e-04  2.459457e-04  3.612092e-04
#>  [3,]  1.565053e-03  1.208249e-03  9.798233e-04  5.993386e-04  8.768969e-04
#>  [4,]  4.496296e-04  3.409994e-04  2.810800e-04  1.700350e-04  2.561577e-04
#>  [5,] -1.027046e-05 -5.195425e-06 -3.616217e-06 -4.622710e-06  4.003915e-06
#>  [6,]  4.663990e-04  3.691991e-04  3.053233e-04  1.820077e-04  2.694452e-04
#>  [7,]  3.664683e-04  2.808658e-04  2.302615e-04  1.383267e-04  2.058062e-04
#>  [8,] -2.915547e-05 -2.357334e-05 -2.066076e-05 -1.381717e-05 -7.607419e-06
#>  [9,] -1.616550e-03 -1.207965e-03 -9.782561e-04 -6.079167e-04 -8.623289e-04
#> [10,]  2.244487e-04  1.741563e-04  1.393139e-04  8.463439e-05  1.333550e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  7.263233e-05  8.191420e-05  6.201008e-05  3.735980e-05  3.602659e-05
#>  [2,]  8.188011e-05  8.943001e-05  6.659981e-05  4.463284e-05  4.255177e-05
#>  [3,]  1.969734e-04  2.218135e-04  1.619695e-04  1.062397e-04  1.033319e-04
#>  [4,]  5.958634e-05  6.452134e-05  4.753620e-05  3.342447e-05  3.220344e-05
#>  [5,]  1.597926e-06 -2.735991e-06 -4.317116e-07  2.341096e-06  1.244540e-06
#>  [6,]  6.228933e-05  6.507457e-05  4.893251e-05  3.506506e-05  3.190895e-05
#>  [7,]  4.764282e-05  5.040574e-05  3.821674e-05  2.669122e-05  2.501768e-05
#>  [8,] -1.127107e-06 -5.428128e-06 -2.473609e-06  7.143306e-07  2.713658e-07
#>  [9,] -1.938354e-04 -2.317026e-04 -1.658952e-04 -9.962048e-05 -1.017908e-04
#> [10,]  3.020225e-05  3.069371e-05  2.395776e-05  1.730237e-05  1.613512e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  2.027431e-05  2.564666e-05  1.011952e-05  3.161203e-05  2.491240e-05
#>  [2,]  2.775185e-05  3.173466e-05  1.715193e-05  3.862757e-05  3.214629e-05
#>  [3,]  6.752174e-05  7.685944e-05  4.010897e-05  9.407384e-05  7.769205e-05
#>  [4,]  2.184346e-05  2.498004e-05  1.478096e-05  3.023940e-05  2.433608e-05
#>  [5,]  1.119970e-06  1.644226e-06  2.893836e-06  8.751764e-07  8.224201e-07
#>  [6,]  2.157431e-05  2.485812e-05  1.553248e-05  2.910573e-05  2.600605e-05
#>  [7,]  1.652456e-05  1.925836e-05  1.142583e-05  2.289859e-05  1.868294e-05
#>  [8,]  2.478612e-07  6.264923e-07  2.160240e-06 -2.857691e-07 -4.909037e-07
#>  [9,] -6.477214e-05 -7.428896e-05 -3.132267e-05 -9.511561e-05 -7.297103e-05
#> [10,]  1.043756e-05  1.223804e-05  7.501042e-06  1.431110e-05  1.169446e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.235300e-04  2.975084e-05  4.424969e-04  4.909738e-04  8.757432e-04
#>  [2,]  3.575486e-04  3.463576e-05  4.808082e-04  5.406266e-04  9.218707e-04
#>  [3,]  8.788075e-04  8.168598e-05  1.170560e-03  1.308834e-03  2.239581e-03
#>  [4,]  2.581615e-04  2.564752e-05  3.449010e-04  3.863035e-04  6.541581e-04
#>  [5,]  3.073402e-06  1.651611e-06  2.955838e-06  9.123181e-06 -1.325697e-06
#>  [6,]  2.738972e-04  2.822700e-05  3.678419e-04  4.204973e-04  6.903424e-04
#>  [7,]  2.067327e-04  2.057444e-05  2.781721e-04  3.130354e-04  5.320004e-04
#>  [8,] -9.617485e-06 -3.624992e-08 -1.609972e-05 -1.351020e-05 -3.572861e-05
#>  [9,] -8.489618e-04 -7.629577e-05 -1.143217e-03 -1.251585e-03 -2.257458e-03
#> [10,]  1.293246e-04  1.306370e-05  1.712759e-04  1.957026e-04  3.278185e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.01306149 -0.03627051 -0.03026683 -0.04951504 -0.05036894 -0.02781386
#> [2,]  0.01306149  0.03627051  0.03026683  0.04951504  0.05036894  0.02781386
#>             [,7]        [,8]       [,9]       [,10]
#> [1,] -0.04681288 -0.04897686 -0.0490531 -0.03512308
#> [2,]  0.04681288  0.04897686  0.0490531  0.03512308
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4302087 0.4242364 0.3504270 0.6281632 0.5083041 0.4344607 0.4399190
#>  [8] 0.3426357 0.3734615 0.2893696
#> 
#> $B[[2]]
#> [1]  0.1017710 -0.1779414
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0016431317  0.0022361090  0.0052832800  0.0017596973  0.0002797367
#>  [6]  0.0019195407  0.0014160046  0.0001863008 -0.0042321055  0.0009424045
#> 
#> $vB[[2]]
#> [1] -0.08870522  0.08870522
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [2,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#>  [3,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#>  [4,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#>  [5,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#>  [6,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [7,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#>  [8,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [9,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [10,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [11,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [12,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [13,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [14,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [15,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [16,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [17,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [18,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [19,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [20,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [21,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [22,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [23,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [24,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [25,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [26,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [27,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [28,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [29,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [30,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [31,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [32,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [33,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [34,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [35,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [36,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [37,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [38,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#> [39,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [40,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [2,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#>  [3,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#>  [4,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#>  [5,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#>  [6,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [7,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#>  [8,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [9,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [10,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [11,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [12,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [13,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [14,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [15,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [16,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [17,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [18,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [19,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [20,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [21,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [22,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [23,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [24,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [25,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [26,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [27,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [28,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [29,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [30,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [31,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [32,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [33,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [34,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [35,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [36,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [37,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [38,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#> [39,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [40,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [2,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#>  [3,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#>  [4,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#>  [5,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#>  [6,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [7,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#>  [8,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [9,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [10,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [11,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [12,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [13,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [14,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [15,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [16,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [17,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [18,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [19,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [20,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [21,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [22,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [23,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [24,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [25,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [26,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [27,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [28,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [29,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [30,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [31,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [32,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [33,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [34,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [35,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [36,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [37,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [38,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#> [39,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [40,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [2,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#>  [3,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#>  [4,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#>  [5,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#>  [6,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [7,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#>  [8,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [9,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [10,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [11,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [12,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [13,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [14,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [15,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [16,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [17,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [18,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [19,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [20,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [21,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [22,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [23,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [24,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [25,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [26,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [27,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [28,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [29,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [30,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [31,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [32,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [33,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [34,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [35,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [36,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [37,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [38,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#> [39,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [40,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [2,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#>  [3,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#>  [4,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#>  [5,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#>  [6,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [7,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#>  [8,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [9,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [10,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [11,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [12,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [13,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [14,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [15,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [16,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [17,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [18,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [19,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [20,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [21,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [22,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [23,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [24,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [25,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [26,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [27,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [28,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [29,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [30,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [31,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [32,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [33,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [34,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [35,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [36,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [37,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [38,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#> [39,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [40,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [2,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#>  [3,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#>  [4,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#>  [5,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#>  [6,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [7,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#>  [8,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [9,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [10,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [11,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [12,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [13,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [14,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [15,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [16,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [17,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [18,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [19,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [20,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [21,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [22,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [23,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [24,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [25,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [26,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [27,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [28,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [29,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [30,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [31,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [32,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [33,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [34,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [35,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [36,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [37,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [38,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#> [39,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [40,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2240787 0.5078296 0.4839517 0.5599569 0.5622209 0.4932865 0.5215273
#>  [2,] 0.2563019 0.5064036 0.4776674 0.5778937 0.5924933 0.4325142 0.5776207
#>  [3,] 0.3645490 0.5644509 0.6276113 0.7816693 0.7442621 0.6216538 0.6889227
#>  [4,] 0.2971636 0.5539554 0.5995931 0.7413782 0.7078720 0.5584607 0.6812464
#>  [5,] 0.2379498 0.5006220 0.5556389 0.6594464 0.6311893 0.4829658 0.6409035
#>  [6,] 0.2698545 0.5077816 0.5396310 0.6355157 0.6057314 0.5190131 0.5552949
#>  [7,] 0.2445903 0.4989261 0.5455577 0.6052848 0.6395394 0.4831488 0.6283229
#>  [8,] 0.3071104 0.5596158 0.6065158 0.7352553 0.7182067 0.5649015 0.6891048
#>  [9,] 0.2907725 0.5497212 0.5823501 0.7052188 0.6787647 0.5118621 0.6734796
#> [10,] 0.3999996 0.5973511 0.7165138 0.8106965 0.7752375 0.6972800 0.7226212
#> [11,] 0.4140105 0.6328714 0.7278833 0.8227943 0.7780381 0.7009314 0.7423205
#> [12,] 0.2755120 0.5546931 0.5900290 0.7151822 0.6725544 0.5401625 0.6775888
#> [13,] 0.3763058 0.5603275 0.6083493 0.7467583 0.7185526 0.6277190 0.6317802
#> [14,] 0.4675549 0.5926555 0.6488949 0.7426108 0.7170065 0.6464999 0.6522045
#> [15,] 0.3009736 0.5317495 0.5644734 0.7119445 0.6654703 0.5068309 0.6371830
#> [16,] 0.3602921 0.5861074 0.5948504 0.7221184 0.6735124 0.5885271 0.6338527
#> [17,] 0.2726358 0.5456266 0.5671313 0.6972910 0.6899807 0.6080146 0.5688893
#> [18,] 0.2440278 0.4911412 0.4507130 0.5328383 0.5949247 0.4113194 0.5747246
#> [19,] 0.2525938 0.5434371 0.5993025 0.7498194 0.6887660 0.5623174 0.6788734
#> [20,] 0.3932787 0.5814617 0.6373728 0.7731457 0.7543462 0.6465769 0.6764693
#> [21,] 0.2901069 0.5658751 0.6227460 0.7571338 0.7088195 0.5710303 0.7094794
#> [22,] 0.3767074 0.6199698 0.6870695 0.8029354 0.7566142 0.6841086 0.7165035
#> [23,] 0.3557976 0.5702752 0.5762820 0.7053022 0.6724969 0.5715784 0.6299811
#> [24,] 0.2494251 0.5388010 0.5637911 0.6775790 0.6555083 0.4836294 0.6645578
#> [25,] 0.3661028 0.5589004 0.5964282 0.7404883 0.7169539 0.6256069 0.6342597
#> [26,] 0.2329233 0.5035434 0.4844327 0.5614569 0.5436985 0.4726202 0.5060847
#> [27,] 0.3683541 0.5876386 0.6846710 0.7702369 0.7335021 0.6629380 0.6808364
#> [28,] 0.3186360 0.5828920 0.5857928 0.7042027 0.6768218 0.5334982 0.6968628
#> [29,] 0.3868751 0.5912764 0.6435595 0.7750782 0.7439157 0.6449894 0.6822089
#> [30,] 0.3869685 0.5518675 0.6176194 0.7480741 0.7238199 0.6173522 0.6520965
#> [31,] 0.2593981 0.5111550 0.4319539 0.5520962 0.6073474 0.4346718 0.5534650
#> [32,] 0.2243778 0.5038912 0.4948637 0.6152895 0.6423739 0.4194521 0.6437867
#> [33,] 0.4282924 0.6268111 0.7003265 0.8057147 0.7612158 0.6992692 0.6830748
#> [34,] 0.3508396 0.5944478 0.6214004 0.6967867 0.6534427 0.6263711 0.6089400
#> [35,] 0.2823083 0.5277793 0.6129538 0.6982186 0.6676196 0.5939188 0.6168478
#> [36,] 0.2877187 0.5187956 0.5191322 0.7031183 0.6732519 0.5285924 0.6080193
#> [37,] 0.3521521 0.5933866 0.6247659 0.7091517 0.6716093 0.6428583 0.6130084
#> [38,] 0.3322044 0.5696606 0.6243902 0.7429804 0.7282742 0.6098438 0.6490787
#> [39,] 0.2991767 0.5953865 0.6474407 0.7851345 0.7507163 0.5808582 0.7462437
#> [40,] 0.2299727 0.5080182 0.4880482 0.6165564 0.6434520 0.4296600 0.6311503
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4641104 0.4570320 0.3605351
#>  [2,] 0.5165600 0.5476558 0.4241057
#>  [3,] 0.6806362 0.7452629 0.5883990
#>  [4,] 0.6651371 0.7135999 0.5482865
#>  [5,] 0.5939267 0.6451816 0.4475575
#>  [6,] 0.5046079 0.5465896 0.4242934
#>  [7,] 0.5672996 0.6091320 0.4508891
#>  [8,] 0.6638600 0.7077417 0.5581708
#>  [9,] 0.6473268 0.6927462 0.5292955
#> [10,] 0.6729573 0.7456181 0.6022987
#> [11,] 0.7036173 0.7703905 0.6304007
#> [12,] 0.6263761 0.6640485 0.4876624
#> [13,] 0.6189195 0.6772325 0.5651760
#> [14,] 0.6408435 0.6938236 0.5996771
#> [15,] 0.6250559 0.6666905 0.5014559
#> [16,] 0.6188540 0.6674702 0.5375054
#> [17,] 0.5472352 0.5780506 0.4835488
#> [18,] 0.5069732 0.5226557 0.4272995
#> [19,] 0.6097519 0.6951969 0.4816458
#> [20,] 0.6662223 0.7188799 0.6066621
#> [21,] 0.6618734 0.7241161 0.5239398
#> [22,] 0.6725901 0.7217484 0.5886809
#> [23,] 0.6192418 0.6536153 0.5446397
#> [24,] 0.5978406 0.6522801 0.4741340
#> [25,] 0.6180954 0.6751529 0.5593874
#> [26,] 0.4643062 0.4790734 0.3543020
#> [27,] 0.6165197 0.6908173 0.5500022
#> [28,] 0.6427827 0.6678990 0.5247701
#> [29,] 0.6660156 0.7184288 0.5991901
#> [30,] 0.6444931 0.6892027 0.5912598
#> [31,] 0.4951149 0.5303188 0.4461971
#> [32,] 0.5868419 0.6193656 0.4446318
#> [33,] 0.6554320 0.7090327 0.6125129
#> [34,] 0.5327864 0.5737333 0.4797488
#> [35,] 0.5551990 0.6406604 0.4598626
#> [36,] 0.5995616 0.6574840 0.4954302
#> [37,] 0.5375743 0.5863452 0.4863617
#> [38,] 0.6158048 0.6702626 0.5478312
#> [39,] 0.6937110 0.7536063 0.5831009
#> [40,] 0.5868135 0.6111104 0.4585319
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4819714 0.5180286
#>  [2,] 0.4857755 0.5142245
#>  [3,] 0.4753913 0.5246087
#>  [4,] 0.4792550 0.5207450
#>  [5,] 0.4844058 0.5155942
#>  [6,] 0.4800719 0.5199281
#>  [7,] 0.4838929 0.5161071
#>  [8,] 0.4780436 0.5219564
#>  [9,] 0.4809104 0.5190896
#> [10,] 0.4670431 0.5329569
#> [11,] 0.4659592 0.5340408
#> [12,] 0.4789306 0.5210694
#> [13,] 0.4739227 0.5260773
#> [14,] 0.4689634 0.5310366
#> [15,] 0.4812010 0.5187990
#> [16,] 0.4752274 0.5247726
#> [17,] 0.4763409 0.5236591
#> [18,] 0.4875776 0.5124224
#> [19,] 0.4799564 0.5200436
#> [20,] 0.4721577 0.5278423
#> [21,] 0.4779188 0.5220812
#> [22,] 0.4679525 0.5320475
#> [23,] 0.4764565 0.5235435
#> [24,] 0.4824617 0.5175383
#> [25,] 0.4748560 0.5251440
#> [26,] 0.4835129 0.5164871
#> [27,] 0.4689385 0.5310615
#> [28,] 0.4774242 0.5225758
#> [29,] 0.4717647 0.5282353
#> [30,] 0.4737149 0.5262851
#> [31,] 0.4874856 0.5125144
#> [32,] 0.4881992 0.5118008
#> [33,] 0.4650756 0.5349244
#> [34,] 0.4695571 0.5304429
#> [35,] 0.4771939 0.5228061
#> [36,] 0.4837626 0.5162374
#> [37,] 0.4694440 0.5305560
#> [38,] 0.4739827 0.5260173
#> [39,] 0.4758491 0.5241509
#> [40,] 0.4877429 0.5122571
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]         [,2]        [,3]      [,4]      [,5]        [,6]
#>  [1,] -1.2420538  0.031320852 -0.06421524 0.2409871 0.2501806 -0.02685565
#>  [2,] -1.0652790  0.025615734 -0.08938998 0.3141327 0.3742823 -0.27160052
#>  [3,] -0.5556739  0.259245740  0.52198267 1.2754207 1.0682402  0.49657355
#>  [4,] -0.8608414  0.216665205  0.40376994 1.0531443 0.8850711  0.23491734
#>  [5,] -1.1639525  0.002487900  0.22348104 0.6608283 0.5373223 -0.06816331
#>  [6,] -0.9953607  0.031128796  0.15885736 0.5559534 0.4294042  0.07608897
#>  [7,] -1.1276759 -0.004295638  0.18273763 0.4275345 0.5733658 -0.06743021
#>  [8,] -0.8136634  0.239603062  0.43268985 1.0214518 0.9355834  0.26107878
#>  [9,] -0.8916352  0.199544468  0.33242811 0.8722745 0.7481008  0.04745717
#> [10,] -0.4054669  0.394440205  0.92723401 1.4545420 1.2381252  0.83437896
#> [11,] -0.3474105  0.544555151  0.98390987 1.5353949 1.2542699  0.85173702
#> [12,] -0.9668338  0.219651337  0.36408526 0.9206876 0.7197610  0.16099668
#> [13,] -0.5052582  0.242491213  0.44037899 1.0813972 0.9372934  0.52244361
#> [14,] -0.1299630  0.374953950  0.61418497 1.0595828 0.9296612  0.60368921
#> [15,] -0.8426658  0.127169256  0.25933750 0.9048469 0.6877684  0.02732532
#> [16,] -0.5740967  0.347896331  0.38405336 0.9549942 0.7241143  0.35787972
#> [17,] -0.9812905  0.183015595  0.27015631 0.8344309 0.8000289  0.43897424
#> [18,] -1.1307225 -0.035439107 -0.19779032 0.1315427 0.3843619 -0.35851382
#> [19,] -1.0848262  0.174187575  0.40255962 1.0976495 0.7943566  0.25057249
#> [20,] -0.4335510  0.328776878  0.56397956 1.2261594 1.1219283  0.60402643
#> [21,] -0.8948649  0.265041132  0.50122010 1.1370294 0.8896574  0.28605593
#> [22,] -0.5035474  0.489419934  0.78645448 1.4047427 1.1342058  0.77271838
#> [23,] -0.5936511  0.282974247  0.30752887 0.8726761 0.7195001  0.28829390
#> [24,] -1.1016809  0.155516744  0.25656269 0.7426678 0.6433411 -0.06550574
#> [25,] -0.5489725  0.236700663  0.39060439 1.0485081 0.9294020  0.51341574
#> [26,] -1.1918775  0.014173691 -0.06228928 0.2470767 0.1752409 -0.10962893
#> [27,] -0.5392843  0.354211980  0.77532182 1.2096493 1.0124643  0.67641479
#> [28,] -0.7600471  0.334656571  0.34659967 0.8673916 0.7392040  0.13419374
#> [29,] -0.4604666  0.369244401  0.59084707 1.2372113 1.0664210  0.59708635
#> [30,] -0.4600730  0.208218949  0.47945589 1.0883671 0.9634896  0.47832488
#> [31,] -1.0490991  0.044627219 -0.27388381 0.2091437 0.4361756 -0.26281500
#> [32,] -1.2403344  0.015565257 -0.02054604 0.4696020 0.5856824 -0.32502304
#> [33,] -0.2888215  0.518560556  0.84885324 1.4224022 1.1593565  0.84382005
#> [34,] -0.6153506  0.382383441  0.49549636 0.8320428 0.6342065  0.51667982
#> [35,] -0.9330405  0.111231608  0.45974561 0.8388293 0.6974386  0.38018932
#> [36,] -0.9064895  0.075217697  0.07656620 0.8621914 0.7229299  0.11449446
#> [37,] -0.6095928  0.377983472  0.50982702 0.8912676 0.7154726  0.58779152
#> [38,] -0.6982315  0.280466424  0.50822483 1.0615172 0.9858844  0.44665579
#> [39,] -0.8512212  0.386278456  0.60780858 1.2958430 1.1024364  0.32629730
#> [40,] -1.2084652  0.032075394 -0.04781614 0.4749572 0.5903784 -0.28323836
#>             [,7]        [,8]        [,9]        [,10]
#>  [1,] 0.08616231 -0.14380580 -0.17229687 -0.573042202
#>  [2,] 0.31301374  0.06626430  0.19120346 -0.305941562
#>  [3,] 0.79508752  0.75669726  1.07350518  0.357350747
#>  [4,] 0.75950549  0.68627210  0.91293257  0.193749710
#>  [5,] 0.57928765  0.38022231  0.59792559 -0.210544391
#>  [6,] 0.22208781  0.01843199  0.18690039 -0.305172878
#>  [7,] 0.52502885  0.27084196  0.44366504 -0.197078902
#>  [8,] 0.79593752  0.68054375  0.88444124  0.233741801
#>  [9,] 0.72396513  0.60730955  0.81298937  0.117316461
#> [10,] 0.95750120  0.72159130  1.07537691  0.415052353
#> [11,] 1.05806464  0.86458324  1.21051733  0.533936152
#> [12,] 0.74271278  0.51670146  0.68138826 -0.049360354
#> [13,] 0.53986136  0.48496446  0.74108247  0.262195770
#> [14,] 0.62874361  0.57902695  0.81805635  0.404119855
#> [15,] 0.56315847  0.51106428  0.69325463  0.005823686
#> [16,] 0.54878071  0.48468707  0.69676521  0.150303773
#> [17,] 0.27732100  0.18950575  0.31477611 -0.065828712
#> [18,] 0.30115405  0.02789455  0.09068492 -0.292877577
#> [19,] 0.74859943  0.44626953  0.82452895 -0.073449773
#> [20,] 0.73759318  0.69114798  0.93891209  0.433302845
#> [21,] 0.89285692  0.67165393  0.96497184  0.095832587
#> [22,] 0.92718318  0.71992310  0.95315070  0.358514945
#> [23,] 0.53213556  0.48633125  0.63496895  0.179035457
#> [24,] 0.68367237  0.39647543  0.62907697 -0.103556324
#> [25,] 0.55053489  0.48147190  0.73158441  0.238676080
#> [26,] 0.02433981 -0.14301867 -0.08375528 -0.600182071
#> [27,] 0.75761833  0.47480202  0.80394315  0.200679624
#> [28,] 0.83240285  0.58746261  0.69869786  0.099161437
#> [29,] 0.76394153  0.69021874  0.93668150  0.402091806
#> [30,] 0.62826723  0.59491940  0.79639447  0.369175964
#> [31,] 0.21468074 -0.01954098  0.12142422 -0.216047978
#> [32,] 0.59183772  0.35092500  0.48685640 -0.222384640
#> [33,] 0.76793844  0.64300295  0.89069070  0.457887393
#> [34,] 0.44285886  0.13133424  0.29709946 -0.081049131
#> [35,] 0.47619014  0.22169969  0.57823174 -0.160895978
#> [36,] 0.43899414  0.40363858  0.65210203 -0.018279545
#> [37,] 0.45997583  0.15058126  0.34887674 -0.054566759
#> [38,] 0.61499188  0.47177929  0.70937306  0.191911459
#> [39,] 1.07867772  0.81752628  1.11793936  0.335516155
#> [40,] 0.53715497  0.35080816  0.45198227 -0.166254380
#> 
#> $pre[[3]]
#>                [,1]       [,2]
#>  [1,] -0.0074115735 0.06473407
#>  [2,]  0.0054039535 0.06231721
#>  [3,]  0.0146769030 0.11319118
#>  [4,]  0.0132273687 0.09625498
#>  [5,]  0.0097096884 0.07210686
#>  [6,] -0.0023637376 0.07739095
#>  [7,]  0.0094275960 0.07387847
#>  [8,]  0.0126205636 0.10050288
#>  [9,]  0.0121806788 0.08857635
#> [10,]  0.0048943531 0.13691324
#> [11,]  0.0053415377 0.14171575
#> [12,]  0.0057942688 0.09012200
#> [13,]  0.0080939467 0.11249791
#> [14,]  0.0059891644 0.13029543
#> [15,]  0.0093099737 0.08454125
#> [16,]  0.0050885849 0.10426030
#> [17,] -0.0012954387 0.09341179
#> [18,]  0.0086722701 0.05837223
#> [19,]  0.0072324717 0.08744981
#> [20,]  0.0110192789 0.12250366
#> [21,]  0.0097354163 0.09811760
#> [22,]  0.0028090690 0.13117497
#> [23,]  0.0076470199 0.10189055
#> [24,]  0.0074147172 0.07759654
#> [25,]  0.0092628296 0.10992366
#> [26,] -0.0063413618 0.05963081
#> [27,] -0.0003781284 0.12402791
#> [28,]  0.0071843752 0.09754921
#> [29,]  0.0091496191 0.12221106
#> [30,]  0.0111554658 0.11639284
#> [31,]  0.0094484320 0.05951636
#> [32,]  0.0131260019 0.06033785
#> [33,] -0.0002880425 0.13963750
#> [34,] -0.0108537638 0.11106875
#> [35,]  0.0014026837 0.09269023
#> [36,]  0.0136421380 0.07861444
#> [37,] -0.0094645509 0.11291198
#> [38,]  0.0043416897 0.10850503
#> [39,]  0.0113336010 0.10801254
#> [40,]  0.0133277140 0.06236575
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5180286 -0.5180286
#>  [2,]  0.5142245 -0.5142245
#>  [3,]  0.5246087 -0.5246087
#>  [4,]  0.5207450 -0.5207450
#>  [5,]  0.5155942 -0.5155942
#>  [6,]  0.5199281 -0.5199281
#>  [7,] -0.4838929  0.4838929
#>  [8,]  0.5219564 -0.5219564
#>  [9,]  0.5190896 -0.5190896
#> [10,] -0.4670431  0.4670431
#> [11,] -0.4659592  0.4659592
#> [12,]  0.5210694 -0.5210694
#> [13,]  0.5260773 -0.5260773
#> [14,] -0.4689634  0.4689634
#> [15,]  0.5187990 -0.5187990
#> [16,] -0.4752274  0.4752274
#> [17,]  0.5236591 -0.5236591
#> [18,] -0.4875776  0.4875776
#> [19,] -0.4799564  0.4799564
#> [20,]  0.5278423 -0.5278423
#> [21,]  0.5220812 -0.5220812
#> [22,] -0.4679525  0.4679525
#> [23,]  0.5235435 -0.5235435
#> [24,]  0.5175383 -0.5175383
#> [25,]  0.5251440 -0.5251440
#> [26,]  0.5164871 -0.5164871
#> [27,] -0.4689385  0.4689385
#> [28,]  0.5225758 -0.5225758
#> [29,] -0.4717647  0.4717647
#> [30,]  0.5262851 -0.5262851
#> [31,]  0.5125144 -0.5125144
#> [32,]  0.5118008 -0.5118008
#> [33,] -0.4650756  0.4650756
#> [34,] -0.4695571  0.4695571
#> [35,] -0.4771939  0.4771939
#> [36,]  0.5162374 -0.5162374
#> [37,] -0.4694440  0.4694440
#> [38,] -0.4739827  0.4739827
#> [39,] -0.4758491  0.4758491
#> [40,]  0.5122571 -0.5122571
#> 
#> $L
#> [1] 0.7132841 0.7102124 0.6913618 0.6999011 0.6969583 0.6965781
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4347826