Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10           V11         V12           V13
#>  [1,] -0.2898650 -0.11172600 -0.0006533508 -0.09035211 -0.0463449422
#>  [2,] -0.3546518 -0.08111954 -0.1092391901 -0.05469518 -0.0792222921
#>  [3,] -0.3605828 -0.04885989 -0.0790770733 -0.07213411 -0.0276632592
#>  [4,] -0.3283792 -0.10797621  0.0770599130 -0.04452897 -0.0212444392
#>  [5,] -0.3242022 -0.15934486 -0.1770392383 -0.02380962 -0.1661500440
#>  [6,] -0.2189091 -0.17777342 -0.1235555668  0.03143806 -0.0003414292
#>  [7,] -0.2292402 -0.04707378 -0.0489825035  0.01256665 -0.0301607582
#>  [8,] -0.2285910 -0.09994546 -0.0342164646 -0.03502108 -0.0788876405
#>  [9,] -0.3331029 -0.08424014 -0.1979553958 -0.21613697 -0.1232449834
#> [10,] -0.2947191 -0.12147406 -0.1311960962 -0.08827970 -0.0836792148
#>                V14          V15         V16         V17          V18
#>  [1,] -0.110028056 -0.052277098  0.03896412 -0.04269542 -0.007839769
#>  [2,] -0.126041743 -0.068839411 -0.07251874 -0.10625597  0.082435460
#>  [3,] -0.052378756 -0.070758851 -0.02936781 -0.14122835  0.036261808
#>  [4,]  0.045322871 -0.067456592  0.13175649  0.19406747  0.077545181
#>  [5,] -0.178025801 -0.120053025 -0.21868595 -0.12334109 -0.091326991
#>  [6,]  0.045949989  0.007314501  0.10446207  0.06600820  0.092126547
#>  [7,]  0.041693523 -0.111046463  0.03413792  0.12593132  0.154207222
#>  [8,] -0.003899314  0.068882326  0.17202049  0.07055401  0.137998896
#>  [9,] -0.171737633 -0.068597568 -0.13185484  0.04111722 -0.023200704
#> [10,] -0.131542307 -0.053308844 -0.10973219  0.09453261  0.088910922
#>               V19         V2          V20          V21         V22          V23
#>  [1,]  0.08269054 -0.3116396  0.163515221  0.190335026  0.14873097  0.175069177
#>  [2,] -0.01109200 -0.2091837  0.080640400  0.075749562  0.04348264  0.039434466
#>  [3,]  0.02729096 -0.3568320 -0.093418965 -0.040931311 -0.05624582  0.052649611
#>  [4,]  0.15734137 -0.3521358  0.178776776  0.230814299  0.26476525  0.185029114
#>  [5,] -0.09881262 -0.2447933 -0.121231052 -0.024719604 -0.04863110 -0.001967111
#>  [6,]  0.03819434 -0.2706865 -0.006108069  0.001572912  0.19549529  0.124023041
#>  [7,]  0.11392100 -0.3024717  0.074951409  0.017289073  0.09747336  0.107353415
#>  [8,]  0.18122870 -0.1757543  0.094101849  0.171357857  0.07887484  0.120451560
#>  [9,] -0.05568662 -0.2710222  0.055727316  0.044104679  0.08844618  0.059303887
#> [10,]  0.06962129 -0.2965971  0.099674028 -0.029458669  0.12539273  0.156703811
#>               V24        V25        V26        V27         V28         V29
#>  [1,]  0.03981952 0.12184756 0.03979486 0.13342474 -0.03781076  0.04747346
#>  [2,]  0.11936385 0.21697570 0.19478549 0.14779060  0.09234781  0.10825649
#>  [3,] -0.02271640 0.14253873 0.08253768 0.08141657  0.02188001  0.08420701
#>  [4,]  0.27730276 0.11029766 0.20620034 0.13894000  0.19660036  0.04541498
#>  [5,]  0.05647498 0.14491467 0.08266197 0.14491455  0.12830807  0.06900399
#>  [6,]  0.06627961 0.16714049 0.23143877 0.17552377  0.21350076  0.16193436
#>  [7,]  0.11158236 0.13611908 0.07769455 0.09925501  0.20262439  0.16402340
#>  [8,]  0.13183960 0.07953160 0.14363927 0.14343390 -0.01498244 -0.04080819
#>  [9,]  0.13955798 0.10720826 0.09640396 0.14366832  0.03791202  0.12898925
#> [10,]  0.18962639 0.08878068 0.12269621 0.04109317  0.09198480  0.07413705
#>               V3         V30         V31         V32         V33          V34
#>  [1,] -0.2996443 -0.05454775 -0.15563849 -0.08424040  0.03559514 -0.005020347
#>  [2,] -0.3059996  0.05659061  0.08732411  0.09707979  0.06922694 -0.083237656
#>  [3,] -0.1823183  0.13820878 -0.07656802 -0.05568564  0.02767239 -0.101563477
#>  [4,] -0.3277550  0.01321323  0.07819997 -0.08818625  0.01918783 -0.113848401
#>  [5,] -0.2221830  0.20442458  0.19972171  0.12924036  0.14644253  0.085274232
#>  [6,] -0.2925237  0.06536169  0.04581561  0.15037124 -0.01204570 -0.013455714
#>  [7,] -0.3056755  0.07486976  0.03895483  0.05694474 -0.04005625 -0.016110861
#>  [8,] -0.3202776 -0.05286714  0.05341452 -0.03925309 -0.05027511 -0.027180427
#>  [9,] -0.3387939  0.05752844  0.07950826 -0.13675588 -0.10995167 -0.025048155
#> [10,] -0.1922826  0.10210999  0.11890234  0.05175397  0.04261168  0.073261726
#>                 V35         V36          V37         V38          V39
#>  [1,]  0.0002116486 -0.09042693 -0.017103076 -0.02860835 -0.043467504
#>  [2,] -0.0351351517 -0.02181816 -0.164128113 -0.05510906 -0.146249837
#>  [3,] -0.0741987669 -0.02995808 -0.052587021 -0.05891766 -0.167195705
#>  [4,] -0.0083706076 -0.15427738 -0.082802510 -0.10296975 -0.096376059
#>  [5,]  0.1180775486  0.07201679 -0.014336657  0.02222670  0.059862897
#>  [6,]  0.0377982146  0.12722688  0.122019868 -0.05053494 -0.001925309
#>  [7,]  0.0230003908 -0.03856187 -0.025852968 -0.07109437 -0.019622126
#>  [8,] -0.0366591091 -0.12128173 -0.122031523 -0.10335870 -0.081943503
#>  [9,] -0.1473862434 -0.18200407  0.005133488 -0.17313436 -0.176361450
#> [10,] -0.0868850675  0.02582851  0.029155334 -0.03496762 -0.086665880
#>               V4         V40          V41          V42         V43          V44
#>  [1,] -0.2281750 -0.13360567 -0.112787380 -0.019011310 -0.04393764 -0.172341714
#>  [2,] -0.3362886 -0.02570987 -0.030679108 -0.033277470 -0.06071450 -0.195254239
#>  [3,] -0.3300213 -0.09216638 -0.185239132 -0.072501650 -0.01947838 -0.036700550
#>  [4,] -0.1766767 -0.14526680 -0.050482960 -0.031502841 -0.16310784 -0.079716474
#>  [5,] -0.2680705 -0.02375104 -0.092792152 -0.124631117 -0.16948961 -0.048840939
#>  [6,] -0.3109199 -0.02796876 -0.036028540  0.046340155  0.04058958  0.009643886
#>  [7,] -0.2291905 -0.04159738 -0.096537159  0.009530343  0.03631393 -0.038529280
#>  [8,] -0.2297754 -0.09070699  0.007942969 -0.094450870 -0.16128664 -0.064887385
#>  [9,] -0.2530027 -0.08129593 -0.101809703 -0.174172146 -0.20281704 -0.130218039
#> [10,] -0.3174011 -0.06184379 -0.089526097 -0.101152798 -0.04650569 -0.063169620
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.08814651 -0.20047406 -0.3200757 -0.3223308 -0.3803457 -0.2618716
#>  [2,] -0.07127310 -0.13289116 -0.1208928 -0.2166580 -0.1973097 -0.3085260
#>  [3,] -0.12393744 -0.23110742 -0.2172492 -0.1276916 -0.2357388 -0.3013328
#>  [4,] -0.01525245 -0.15745997 -0.1277725 -0.1246454 -0.2574661 -0.1740072
#>  [5,] -0.07527679 -0.16927936 -0.1344202 -0.1199294 -0.2580113 -0.2610398
#>  [6,] -0.17447078 -0.19941868 -0.1520970 -0.2412091 -0.1936836 -0.2172209
#>  [7,] -0.13006688 -0.07583180 -0.2160255 -0.2971252 -0.3341022 -0.1563231
#>  [8,] -0.10280597 -0.07570473 -0.1474446 -0.1061741 -0.2893959 -0.1318072
#>  [9,] -0.09942378 -0.14664974 -0.1239739 -0.1145714 -0.3115315 -0.1953111
#> [10,] -0.15685280 -0.11553862 -0.0678310 -0.2687849 -0.1989404 -0.3027089
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3632989 -0.4439895 -0.2867050 -0.3093549 -0.4005026 -0.3023713
#>  [2,] -0.3114607 -0.2432739 -0.3224513 -0.3492009 -0.2408606 -0.4461766
#>  [3,] -0.3752549 -0.4138423 -0.4327434 -0.2569786 -0.3565712 -0.3847545
#>  [4,] -0.2610982 -0.2457495 -0.2321094 -0.2771914 -0.3658533 -0.3998353
#>  [5,] -0.3360637 -0.2056831 -0.3501072 -0.2629667 -0.2624283 -0.2888517
#>  [6,] -0.2842734 -0.2345491 -0.3287979 -0.3584792 -0.2908614 -0.3999278
#>  [7,] -0.3811984 -0.4173131 -0.4348672 -0.4476259 -0.4218212 -0.2987682
#>  [8,] -0.2692831 -0.3690258 -0.3376309 -0.2617309 -0.3806809 -0.3017523
#>  [9,] -0.2865204 -0.2171621 -0.3532521 -0.2585801 -0.2390472 -0.2641280
#> [10,] -0.3557930 -0.3648370 -0.3084915 -0.2702949 -0.3258928 -0.2930580
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3294324 -0.3944248 -0.4128152 -0.3159894 -0.12824625 -0.3797635
#>  [2,] -0.3631548 -0.3078884 -0.2689390 -0.3134090 -0.25661539 -0.2724683
#>  [3,] -0.2691750 -0.3029944 -0.4052349 -0.3964118 -0.29649111 -0.4086985
#>  [4,] -0.3381953 -0.3967303 -0.2421292 -0.2772107 -0.08495367 -0.2766284
#>  [5,] -0.2981441 -0.2785271 -0.3093125 -0.2331001 -0.21549536 -0.2278516
#>  [6,] -0.3316006 -0.4494917 -0.3674722 -0.3336198 -0.26233474 -0.3236600
#>  [7,] -0.3093822 -0.2697615 -0.3232655 -0.4166767 -0.08164389 -0.3786079
#>  [8,] -0.3225362 -0.2685315 -0.3894949 -0.4074296 -0.25085925 -0.3268718
#>  [9,] -0.3186251 -0.2232694 -0.3257212 -0.3116127 -0.11144777 -0.3090341
#> [10,] -0.3955703 -0.2900149 -0.2667375 -0.2797717 -0.16146713 -0.3994142
#>                V7          V8           V9
#>  [1,] -0.17188584 -0.10762110 -0.228972413
#>  [2,] -0.20262693 -0.25367346 -0.093270745
#>  [3,] -0.09709285 -0.17155743 -0.122971539
#>  [4,] -0.10810933 -0.07811693 -0.006918131
#>  [5,] -0.25798568 -0.25104748 -0.069116097
#>  [6,] -0.12760246 -0.14588321 -0.190138490
#>  [7,] -0.15237103 -0.12927989 -0.160399043
#>  [8,] -0.15255775 -0.19805695 -0.035853085
#>  [9,] -0.10499511 -0.24463270 -0.130537286
#> [10,] -0.21503710 -0.18003890 -0.184128577
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,]  0.09120800 -0.04796268 -0.100951650  0.13993060 -0.04239656 -0.01340521
#> [2,] -0.05155122 -0.08359175 -0.006072444 -0.05624485  0.01173344 -0.10590380
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.02166806  0.12244252 -0.09514502 -0.03080384
#> [2,] -0.06653402 -0.02455822  0.09038210 -0.02577817
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.311001e-04 -0.0011663204 -0.0014980920 -1.255967e-03 -1.001179e-03
#>  [2,]  4.397167e-05  0.0002595573  0.0003341738  4.178025e-04  4.461140e-04
#>  [3,]  1.780590e-04  0.0014498898  0.0018596320  1.807102e-03  1.602993e-03
#>  [4,] -8.267845e-05 -0.0006617673 -0.0008381976 -6.071330e-04 -4.648280e-04
#>  [5,]  1.057828e-04  0.0007159581  0.0009181249  9.720041e-04  9.274741e-04
#>  [6,] -1.399758e-05 -0.0001345285 -0.0001633833 -5.249704e-05  1.272355e-05
#>  [7,]  5.320683e-05  0.0003276184  0.0004208987  4.750630e-04  4.904047e-04
#>  [8,] -8.608736e-05 -0.0007609704 -0.0009725475 -7.630711e-04 -5.848932e-04
#>  [9,]  3.183471e-04  0.0026383165  0.0034088314  3.266473e-03  2.853162e-03
#> [10,]  9.552900e-05  0.0006386953  0.0008212443  8.625371e-04  8.389489e-04
#>                 V14           V15           V16          V17          V18
#>  [1,] -6.356731e-04 -0.0003891360 -1.854534e-05 0.0000580674 7.031203e-05
#>  [2,]  4.736190e-04  0.0004587233  4.937008e-04 0.0006512585 7.396643e-04
#>  [3,]  1.308260e-03  0.0010476684  7.857435e-04 0.0009857133 1.153604e-03
#>  [4,] -2.772852e-04 -0.0001668713  9.887154e-05 0.0001403230 2.551558e-04
#>  [5,]  8.479381e-04  0.0007465469  6.697594e-04 0.0009170751 1.044722e-03
#>  [6,]  7.745944e-05  0.0001060728  1.973154e-04 0.0002811110 3.501224e-04
#>  [7,]  4.941029e-04  0.0004674126  4.737448e-04 0.0006063933 6.777452e-04
#>  [8,] -3.298216e-04 -0.0001712869  1.065946e-04 0.0001799464 2.437938e-04
#>  [9,]  2.249618e-03  0.0017478503  1.209711e-03 0.0014923941 1.756838e-03
#> [10,]  7.894789e-04  0.0007108041  6.526144e-04 0.0008409506 9.367746e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -6.569082e-04 -1.026832e-04 -0.0011144342 -0.0010775615 -0.0011445493
#>  [2,]  7.550620e-04  5.193928e-05  0.0008380062  0.0009335423  0.0009266121
#>  [3,]  1.814156e-03  1.695428e-04  0.0023118412  0.0024634901  0.0025923050
#>  [4,] -9.011652e-05 -5.835923e-05 -0.0004403508 -0.0004082803 -0.0003798006
#>  [5,]  1.256816e-03  1.082953e-04  0.0015356698  0.0016864816  0.0016814977
#>  [6,]  2.154858e-04 -4.042866e-06  0.0001178739  0.0001594534  0.0001667619
#>  [7,]  7.442383e-04  5.863064e-05  0.0008594168  0.0009436351  0.0009370656
#>  [8,] -2.219721e-04 -6.241393e-05 -0.0005533663 -0.0005171817 -0.0005455025
#>  [9,]  3.070940e-03  2.933429e-04  0.0039612338  0.0041592055  0.0044074731
#> [10,]  1.133907e-03  9.781056e-05  0.0013773394  0.0014919504  0.0014890676
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0010085544 -6.262254e-04 -0.0002704874 -5.286892e-04 -5.886400e-04
#>  [2,]  0.0010224409  1.043222e-03  0.0010746949  1.217208e-03  1.273734e-03
#>  [3,]  0.0027037240  2.430544e-03  0.0021892479  2.652710e-03  2.836434e-03
#>  [4,] -0.0002224974  2.042387e-05  0.0001326593 -2.768272e-05  1.837493e-06
#>  [5,]  0.0018154643  1.769295e-03  0.0017383911  2.057128e-03  2.174133e-03
#>  [6,]  0.0002595984  3.522022e-04  0.0004355529  4.419561e-04  4.774701e-04
#>  [7,]  0.0010077415  9.975215e-04  0.0009986385  1.139060e-03  1.192778e-03
#>  [8,] -0.0004146774 -1.546579e-04  0.0000404189 -1.421433e-04 -1.509858e-04
#>  [9,]  0.0045335862  3.948121e-03  0.0034062967  4.189763e-03  4.510286e-03
#> [10,]  0.0015767720  1.509304e-03  0.0014652802  1.721597e-03  1.807593e-03
#>                 V28           V29            V3           V30          V31
#>  [1,] -6.179209e-04 -4.522774e-04 -8.433014e-05 -1.976987e-04 0.0003161059
#>  [2,]  1.140808e-03  9.483790e-04  5.143331e-05  7.123482e-04 0.0006854584
#>  [3,]  2.618623e-03  2.088527e-03  1.579134e-04  1.424767e-03 0.0009645003
#>  [4,]  2.443606e-05  1.508097e-04 -2.962450e-05  2.077717e-04 0.0004829451
#>  [5,]  1.972123e-03  1.604876e-03  9.731329e-05  1.146863e-03 0.0009818135
#>  [6,]  4.212282e-04  3.613590e-04  3.378103e-06  2.712193e-04 0.0003739057
#>  [7,]  1.087341e-03  9.176105e-04  5.669809e-05  6.925022e-04 0.0006149587
#>  [8,] -1.674362e-04 -5.972117e-05 -4.246049e-05  6.249355e-05 0.0003941369
#>  [9,]  4.194599e-03  3.354939e-03  2.760889e-04  2.259192e-03 0.0013649073
#> [10,]  1.644430e-03  1.364473e-03  9.013608e-05  1.001493e-03 0.0008256877
#>                 V32          V33           V34           V35           V36
#>  [1,] -3.814815e-05 0.0001213636  0.0008398713  0.0014205385  1.709213e-03
#>  [2,]  6.716325e-04 0.0006194550  0.0005455827  0.0004466352  3.193466e-04
#>  [3,]  1.215740e-03 0.0009298590  0.0001457125 -0.0005158027 -9.705070e-04
#>  [4,]  1.715598e-04 0.0002750940  0.0006997373  0.0010165649  1.234046e-03
#>  [5,]  1.080813e-03 0.0009497378  0.0006358170  0.0003209670  3.783273e-05
#>  [6,]  2.766420e-04 0.0002735939  0.0003680470  0.0004231636  4.461391e-04
#>  [7,]  6.207223e-04 0.0005690667  0.0004570866  0.0003314606  1.979542e-04
#>  [8,]  1.109205e-04 0.0002172597  0.0007105442  0.0011014448  1.297176e-03
#>  [9,]  1.869799e-03 0.0013660018 -0.0001019369 -0.0013358352 -2.138488e-03
#> [10,]  9.008057e-04 0.0008008522  0.0005284308  0.0002574395  1.665887e-05
#>                 V37           V38          V39            V4           V40
#>  [1,]  1.500123e-03  0.0008550374 4.791366e-04 -1.915064e-04  6.902785e-04
#>  [2,]  3.136923e-04  0.0003237360 3.361110e-04  8.229986e-05  3.723529e-04
#>  [3,] -7.627801e-04 -0.0002201454 1.284947e-04  3.007748e-04  3.803513e-05
#>  [4,]  1.152855e-03  0.0007786635 6.228737e-04 -1.007269e-04  7.343244e-04
#>  [5,]  6.848384e-05  0.0002366057 3.444966e-04  1.710660e-04  3.527381e-04
#>  [6,]  4.237820e-04  0.0003058972 2.624030e-04 -4.722992e-06  3.409384e-04
#>  [7,]  1.982454e-04  0.0002590591 2.954014e-04  9.190243e-05  3.019684e-04
#>  [8,]  1.174271e-03  0.0007364718 4.909391e-04 -1.145109e-04  6.312920e-04
#>  [9,] -1.733926e-03 -0.0006368751 5.591214e-05  5.318744e-04 -1.686700e-04
#> [10,]  4.676235e-05  0.0002190588 3.193333e-04  1.578949e-04  3.000711e-04
#>                V41           V42           V43           V44           V45
#>  [1,] 0.0001315049 -5.971952e-04 -7.524769e-04 -6.948322e-04 -1.034113e-03
#>  [2,] 0.0003757798  3.356208e-04  2.833779e-04  1.460435e-04  8.227882e-05
#>  [3,] 0.0005145474  1.056647e-03  1.129137e-03  8.541436e-04  9.860143e-04
#>  [4,] 0.0002974006 -1.538783e-04 -3.085745e-04 -3.358380e-04 -6.208154e-04
#>  [5,] 0.0005175943  6.644027e-04  6.413320e-04  4.029002e-04  3.909673e-04
#>  [6,] 0.0002227565  4.079709e-05 -2.147776e-05 -7.930317e-05 -1.932739e-04
#>  [7,] 0.0003487839  3.670746e-04  3.259280e-04  1.979801e-04  1.765527e-04
#>  [8,] 0.0002072604 -3.135906e-04 -4.511670e-04 -4.534408e-04 -7.318695e-04
#>  [9,] 0.0007709789  1.900277e-03  2.049157e-03  1.596575e-03  1.916744e-03
#> [10,] 0.0004549799  5.993987e-04  5.698453e-04  3.742412e-04  3.913223e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0009352046 -5.213950e-04 -4.691593e-04 -2.872698e-04 -2.651554e-04
#>  [2,]  0.0001374087  1.231929e-04  7.884283e-05  5.631472e-05  9.062982e-05
#>  [3,]  0.0009949429  6.431469e-04  5.200878e-04  3.286344e-04  3.802449e-04
#>  [4,] -0.0005753137 -2.805186e-04 -2.557186e-04 -1.580271e-04 -1.158165e-04
#>  [5,]  0.0004696365  3.244910e-04  2.388115e-04  1.604811e-04  1.952778e-04
#>  [6,] -0.0001451022 -4.646723e-05 -5.843063e-05 -3.074957e-05 -1.006288e-05
#>  [7,]  0.0002146209  1.649625e-04  1.189556e-04  7.917392e-05  1.050018e-04
#>  [8,] -0.0006561485 -3.470993e-04 -3.156396e-04 -1.923393e-04 -1.545143e-04
#>  [9,]  0.0018831154  1.195337e-03  9.798830e-04  6.160238e-04  6.886127e-04
#> [10,]  0.0004382044  2.994213e-04  2.272664e-04  1.493404e-04  1.866093e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -4.439513e-05 -8.839879e-05 -8.505855e-05 -1.814921e-05 -5.207951e-06
#>  [2,]  3.014583e-05  1.456805e-05  1.635832e-05  1.104620e-05  1.039326e-05
#>  [3,]  8.175073e-05  9.607848e-05  9.797247e-05  3.253045e-05  1.982690e-05
#>  [4,] -2.326858e-05 -4.940917e-05 -5.214455e-05 -4.801228e-06  5.849409e-06
#>  [5,]  5.724656e-05  4.467917e-05  4.849144e-05  1.951375e-05  1.555968e-05
#>  [6,]  3.466986e-06 -1.214920e-05 -1.202889e-05  5.001850e-07  3.235793e-06
#>  [7,]  3.310472e-05  2.277472e-05  2.257367e-05  1.273008e-05  1.146715e-05
#>  [8,] -2.642880e-05 -5.916761e-05 -5.777320e-05 -8.214556e-06  1.874446e-06
#>  [9,]  1.422821e-04  1.807571e-04  1.813487e-04  5.712958e-05  3.311416e-05
#> [10,]  5.251362e-05  4.372189e-05  4.491180e-05  1.982057e-05  1.593348e-05
#>                 V55          V56          V57           V58           V59
#>  [1,] -2.730371e-06 5.931313e-06 5.732592e-07 -2.108915e-05 -4.870925e-07
#>  [2,]  1.357660e-05 6.693040e-06 1.090963e-05  7.589735e-06  6.662860e-06
#>  [3,]  2.172685e-05 5.625306e-06 1.698359e-05  3.030663e-05  1.015083e-05
#>  [4,]  1.824642e-06 8.803550e-06 1.955738e-06 -5.982012e-06  6.678567e-06
#>  [5,]  2.074653e-05 6.209107e-06 1.622671e-05  1.701954e-05  8.876124e-06
#>  [6,]  4.979886e-06 3.764115e-06 4.320421e-06 -1.256384e-06  2.603573e-06
#>  [7,]  1.393295e-05 6.463873e-06 1.086499e-05  9.207315e-06  7.080482e-06
#>  [8,]  1.808962e-06 7.768737e-06 2.933871e-06 -1.051076e-05  3.545406e-06
#>  [9,]  3.389230e-05 7.062138e-06 2.566509e-05  5.536939e-05  1.729198e-05
#> [10,]  1.968713e-05 7.099712e-06 1.468991e-05  1.565673e-05  9.260018e-06
#>                  V6          V60            V7            V8            V9
#>  [1,] -1.128648e-04 7.504611e-06 -8.973449e-05 -3.320758e-04 -8.116727e-04
#>  [2,]  1.663444e-04 5.969462e-06  1.940012e-04  1.995277e-04  2.018985e-04
#>  [3,]  3.765984e-04 3.161572e-06  4.011567e-04  6.260277e-04  1.044092e-03
#>  [4,] -2.242388e-05 9.361976e-06 -1.485778e-05 -1.620683e-04 -4.772162e-04
#>  [5,]  2.795085e-04 5.549293e-06  3.143034e-04  3.922218e-04  5.301310e-04
#>  [6,]  5.528344e-05 4.243230e-06  7.182735e-05  1.965749e-05 -9.440299e-05
#>  [7,]  1.630961e-04 5.605690e-06  1.894450e-04  2.136463e-04  2.503562e-04
#>  [8,] -3.960405e-05 8.024802e-06 -2.032535e-05 -1.860571e-04 -5.283809e-04
#>  [9,]  6.163016e-04 2.160352e-06  6.446594e-04  1.078237e-03  1.889198e-03
#> [10,]  2.467529e-04 6.020421e-06  2.777879e-04  3.489296e-04  4.774133e-04
#> 
#> $vW[[2]]
#>             [,1]       [,2]        [,3]        [,4]        [,5]      [,6]
#> [1,] -0.02675285 -0.0321247 -0.01764712 -0.05225917 -0.01369495 -0.040606
#> [2,]  0.02675285  0.0321247  0.01764712  0.05225917  0.01369495  0.040606
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.04024399 -0.03871256 -0.02204584 -0.03000208
#> [2,]  0.04024399  0.03871256  0.02204584  0.03000208
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.2734655 0.3314103 0.1844772 0.4788152 0.3520705 0.7171502 0.5259348
#>  [8] 0.4160974 0.2162301 0.4155080
#> 
#> $B[[2]]
#> [1] -0.009580933 -0.049451827
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 1.360486e-05 1.504376e-03 2.561749e-03 5.915155e-04 2.291271e-03
#>  [6] 6.902600e-04 1.401579e-03 4.012442e-04 3.899029e-03 1.977716e-03
#> 
#> $vB[[2]]
#> [1] -0.05365438  0.05365438
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#>  [2,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#>  [3,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#>  [4,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#>  [5,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#>  [6,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#>  [7,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#>  [8,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [9,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [10,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [11,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [12,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [13,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [14,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [15,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [16,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [17,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [18,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [19,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [20,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [21,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [22,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [23,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [24,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [25,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [26,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [27,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [28,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [29,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [30,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [31,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [32,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [33,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [34,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [35,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [36,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [37,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [38,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [39,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [40,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#>  [2,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#>  [3,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#>  [4,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#>  [5,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#>  [6,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#>  [7,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#>  [8,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [9,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [10,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [11,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [12,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [13,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [14,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [15,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [16,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [17,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [18,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [19,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [20,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [21,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [22,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [23,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [24,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [25,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [26,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [27,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [28,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [29,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [30,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [31,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [32,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [33,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [34,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [35,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [36,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [37,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [38,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [39,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [40,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#>  [2,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#>  [3,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#>  [4,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#>  [5,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#>  [6,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#>  [7,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#>  [8,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [9,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [10,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [11,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [12,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [13,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [14,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [15,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [16,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [17,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [18,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [19,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [20,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [21,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [22,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [23,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [24,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [25,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [26,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [27,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [28,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [29,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [30,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [31,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [32,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [33,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [34,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [35,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [36,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [37,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [38,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [39,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [40,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#>  [2,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#>  [3,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#>  [4,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#>  [5,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#>  [6,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#>  [7,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#>  [8,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [9,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [10,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [11,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [12,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [13,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [14,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [15,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [16,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [17,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [18,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [19,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [20,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [21,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [22,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [23,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [24,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [25,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [26,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [27,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [28,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [29,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [30,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [31,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [32,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [33,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [34,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [35,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [36,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [37,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [38,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [39,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [40,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#>  [2,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#>  [3,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#>  [4,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#>  [5,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#>  [6,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#>  [7,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#>  [8,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [9,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [10,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [11,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [12,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [13,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [14,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [15,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [16,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [17,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [18,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [19,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [20,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [21,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [22,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [23,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [24,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [25,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [26,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [27,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [28,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [29,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [30,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [31,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [32,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [33,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [34,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [35,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [36,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [37,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [38,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [39,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [40,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#>  [2,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#>  [3,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#>  [4,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#>  [5,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#>  [6,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#>  [7,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#>  [8,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [9,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [10,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [11,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [12,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [13,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [14,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [15,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [16,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [17,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [18,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [19,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [20,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [21,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [22,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [23,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [24,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [25,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [26,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [27,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [28,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [29,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [30,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [31,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [32,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [33,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [34,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [35,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [36,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [37,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [38,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [39,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [40,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4903181 0.5752052 0.4475138 0.6596619 0.6466254 0.7731966 0.6964083
#>  [2,] 0.3974283 0.4059059 0.2600207 0.5939432 0.4792539 0.6660564 0.5610750
#>  [3,] 0.3124393 0.4624105 0.2771755 0.6026957 0.4825351 0.7243759 0.6098396
#>  [4,] 0.5560855 0.6808914 0.4536562 0.8115548 0.6296447 0.8449505 0.7622343
#>  [5,] 0.5300895 0.5693393 0.4099474 0.7333448 0.6170796 0.8213609 0.7193043
#>  [6,] 0.5768445 0.7068751 0.5008592 0.8122707 0.6833157 0.8590577 0.7809311
#>  [7,] 0.5076268 0.5521285 0.3390820 0.7705917 0.5086107 0.8146032 0.6950729
#>  [8,] 0.5557558 0.5037850 0.3384030 0.7849293 0.3629167 0.7678312 0.7084808
#>  [9,] 0.5975945 0.6412476 0.4453652 0.7940583 0.6551028 0.8768098 0.7739228
#> [10,] 0.5425384 0.6808803 0.4869174 0.8200879 0.6461816 0.8390804 0.7652304
#> [11,] 0.4341970 0.4778606 0.3031876 0.7778226 0.4174560 0.7812401 0.7174034
#> [12,] 0.4304670 0.5289761 0.3320880 0.6277866 0.5861796 0.7583652 0.6348953
#> [13,] 0.5031846 0.6426096 0.4235932 0.7607772 0.7286956 0.8902896 0.7805793
#> [14,] 0.5348417 0.5605101 0.4061519 0.7733799 0.4698523 0.7770582 0.6968457
#> [15,] 0.5957108 0.7183025 0.5313395 0.8445904 0.6714069 0.8621788 0.7971924
#> [16,] 0.6268294 0.6919187 0.4586102 0.8983472 0.5207136 0.8783616 0.8208789
#> [17,] 0.4674012 0.5438473 0.4010300 0.7023865 0.5223081 0.7590920 0.6799896
#> [18,] 0.5441412 0.6855133 0.4874648 0.8039513 0.6491228 0.8422437 0.7706743
#> [19,] 0.4062298 0.6072626 0.3993588 0.7001824 0.6891795 0.8402933 0.7174647
#> [20,] 0.4880995 0.6068100 0.3270135 0.7727919 0.5927773 0.8077308 0.7293836
#> [21,] 0.6226826 0.6282805 0.4405055 0.8764866 0.4915818 0.8512826 0.8097311
#> [22,] 0.5389595 0.6430459 0.4321422 0.7812687 0.6293742 0.8358917 0.7599984
#> [23,] 0.4302959 0.5275087 0.3793250 0.5600429 0.6846398 0.7940797 0.6564534
#> [24,] 0.5237111 0.5870422 0.4544882 0.6796383 0.6736268 0.7882629 0.7104729
#> [25,] 0.5880502 0.6945941 0.4814204 0.7974795 0.7321509 0.8735182 0.7790540
#> [26,] 0.5685739 0.7333680 0.5292461 0.8317499 0.7140167 0.8635401 0.7878252
#> [27,] 0.5058177 0.6156349 0.3842547 0.8099006 0.5517658 0.8440256 0.7513067
#> [28,] 0.5776537 0.6743720 0.4859188 0.8388791 0.5873825 0.8631948 0.7853398
#> [29,] 0.5243176 0.6603003 0.4765695 0.7591886 0.7108427 0.8534176 0.7671018
#> [30,] 0.5803132 0.6229963 0.4366268 0.8279520 0.5625082 0.8452065 0.7936875
#> [31,] 0.5493600 0.5832694 0.4118532 0.8272702 0.4601209 0.8336354 0.7520506
#> [32,] 0.5842755 0.6198356 0.4127579 0.7888430 0.5677210 0.8202952 0.7381310
#> [33,] 0.6043209 0.7252625 0.5314767 0.8417458 0.6778901 0.8737626 0.8015684
#> [34,] 0.4566367 0.5794716 0.4113168 0.6228349 0.6968221 0.7872156 0.6750357
#> [35,] 0.5838829 0.6917198 0.5341718 0.7723260 0.6544792 0.8248888 0.7416784
#> [36,] 0.5343381 0.4525634 0.3295829 0.8016936 0.2863953 0.7501879 0.7047236
#> [37,] 0.6186976 0.7015859 0.4968228 0.8729415 0.5818029 0.8484595 0.7986101
#> [38,] 0.5946413 0.6460978 0.4306060 0.8688775 0.4944734 0.8430250 0.7827405
#> [39,] 0.6452247 0.7323933 0.5185208 0.8780233 0.6353292 0.8655783 0.8108061
#> [40,] 0.3625177 0.4767817 0.2983012 0.6114130 0.5536671 0.7772724 0.6442047
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5536714 0.4543019 0.6421681
#>  [2,] 0.4773544 0.2589695 0.4863599
#>  [3,] 0.4605541 0.2666416 0.5068144
#>  [4,] 0.6731673 0.4953791 0.7027285
#>  [5,] 0.5773631 0.4015935 0.6395390
#>  [6,] 0.6730263 0.5467191 0.7182497
#>  [7,] 0.6390458 0.4031248 0.6252792
#>  [8,] 0.6958884 0.3452844 0.6064478
#>  [9,] 0.6217263 0.4483343 0.7090421
#> [10,] 0.6590966 0.5332790 0.6956346
#> [11,] 0.6094858 0.2990626 0.5585257
#> [12,] 0.5360327 0.3239454 0.5843207
#> [13,] 0.6007321 0.4309006 0.7332647
#> [14,] 0.6390061 0.4462000 0.5920356
#> [15,] 0.6886438 0.5805566 0.7297311
#> [16,] 0.7672604 0.5500867 0.7324132
#> [17,] 0.6034668 0.3963617 0.5868850
#> [18,] 0.6544434 0.5425605 0.6915552
#> [19,] 0.5261768 0.3806985 0.6442507
#> [20,] 0.6108941 0.3509642 0.6632569
#> [21,] 0.7508157 0.5154374 0.7052163
#> [22,] 0.6325228 0.4390234 0.6646848
#> [23,] 0.4713112 0.3406710 0.6064275
#> [24,] 0.5677465 0.4691615 0.6599252
#> [25,] 0.6464200 0.5225815 0.7410494
#> [26,] 0.6773619 0.5879568 0.7395516
#> [27,] 0.6220388 0.3998508 0.6390637
#> [28,] 0.7153970 0.5249012 0.7081451
#> [29,] 0.6110036 0.4919448 0.7164957
#> [30,] 0.6934092 0.4771899 0.7134429
#> [31,] 0.7083098 0.4502929 0.6660992
#> [32,] 0.6749616 0.4597467 0.6699584
#> [33,] 0.6983849 0.5807018 0.7374499
#> [34,] 0.5281490 0.3913689 0.6599638
#> [35,] 0.6685966 0.5696160 0.6832695
#> [36,] 0.6881011 0.3441342 0.5465405
#> [37,] 0.7239140 0.5809878 0.7173864
#> [38,] 0.7335216 0.4980266 0.6810436
#> [39,] 0.7370635 0.5982759 0.7441223
#> [40,] 0.4780995 0.2845174 0.5587494
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4466347 0.5533653
#>  [2,] 0.4669459 0.5330541
#>  [3,] 0.4629057 0.5370943
#>  [4,] 0.4492957 0.5507043
#>  [5,] 0.4544126 0.5455874
#>  [6,] 0.4436957 0.5563043
#>  [7,] 0.4607722 0.5392278
#>  [8,] 0.4698316 0.5301684
#>  [9,] 0.4515225 0.5484775
#> [10,] 0.4454642 0.5545358
#> [11,] 0.4690672 0.5309328
#> [12,] 0.4584028 0.5415972
#> [13,] 0.4481076 0.5518924
#> [14,] 0.4579547 0.5420453
#> [15,] 0.4420545 0.5579455
#> [16,] 0.4523982 0.5476018
#> [17,] 0.4564768 0.5435232
#> [18,] 0.4444448 0.5555552
#> [19,] 0.4498376 0.5501624
#> [20,] 0.4602387 0.5397613
#> [21,] 0.4555838 0.5444162
#> [22,] 0.4523075 0.5476925
#> [23,] 0.4508398 0.5491602
#> [24,] 0.4460033 0.5539967
#> [25,] 0.4439386 0.5560614
#> [26,] 0.4395883 0.5604117
#> [27,] 0.4580966 0.5419034
#> [28,] 0.4490610 0.5509390
#> [29,] 0.4440666 0.5559334
#> [30,] 0.4534517 0.5465483
#> [31,] 0.4587920 0.5412080
#> [32,] 0.4551347 0.5448653
#> [33,] 0.4420248 0.5579752
#> [34,] 0.4482549 0.5517451
#> [35,] 0.4420023 0.5579977
#> [36,] 0.4728292 0.5271708
#> [37,] 0.4469630 0.5530370
#> [38,] 0.4562232 0.5437768
#> [39,] 0.4444255 0.5555745
#> [40,] 0.4602764 0.5397236
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]        [,3]      [,4]        [,5]      [,6]
#>  [1,] -0.03873226  0.30312045 -0.21072116 0.6617878  0.60423849 1.2264499
#>  [2,] -0.41619216 -0.38091637 -1.04586095 0.3802905 -0.08303218 0.6904022
#>  [3,] -0.78873975 -0.15064218 -0.95851546 0.4167101 -0.06988800 0.9662727
#>  [4,]  0.22529003  0.75787140 -0.18590891 1.4601444  0.53069284 1.6955338
#>  [5,]  0.12050379  0.27915602 -0.36418288 1.0116595  0.47717070 1.5255947
#>  [6,]  0.30983320  0.88025498  0.00343670 1.4648329  0.76905151 1.8074855
#>  [7,]  0.03050942  0.20927460 -0.66738785 1.2116553  0.03444621 1.4802026
#>  [8,]  0.22395472  0.01514047 -0.67041924 1.2946269 -0.56272730 1.1961052
#>  [9,]  0.39545210  0.58078328 -0.21941529 1.3495640  0.64154577 1.9625607
#> [10,]  0.17056611  0.75782034 -0.05234248 1.5169428  0.60229679 1.6514016
#> [11,] -0.26474769 -0.08861561 -0.83216459 1.2530222 -0.33322537 1.2729076
#> [12,] -0.27994599  0.11603456 -0.69875654 0.5227329  0.34819436 1.1437375
#> [13,]  0.01273861  0.58670879 -0.30804020 1.1569452  0.98801461 2.0937029
#> [14,]  0.13959323  0.24323243 -0.37989622 1.2274952 -0.12073723 1.2486043
#> [15,]  0.38762470  0.93605693  0.12552265 1.6927875  0.71455534 1.8335058
#> [16,]  0.51863864  0.80910452 -0.16593881 2.1789933  0.08290176 1.9770061
#> [17,] -0.13058056  0.17584075 -0.40117546 0.8586880  0.08929182 1.1477079
#> [18,]  0.17702553  0.77922613 -0.05015121 1.4111756  0.61518549 1.6750181
#> [19,] -0.37957326  0.43582023 -0.40813767 0.8481664  0.79628600 1.6604118
#> [20,] -0.04761099  0.43392284 -0.72172368 1.2241437  0.37545871 1.4353324
#> [21,]  0.50095034  0.52484726 -0.23911086 1.9595714 -0.03367597 1.7446966
#> [22,]  0.15615473  0.58860872 -0.27311649 1.2730753  0.52953290 1.6279725
#> [23,] -0.28064379  0.11014604 -0.49241437 0.2413360  0.77517746 1.3496946
#> [24,]  0.09491574  0.35175147 -0.18255267 0.7521102  0.72463466 1.3144861
#> [25,]  0.35591095  0.82168600 -0.07435265 1.3706148  1.00556280 1.9324307
#> [26,]  0.27603510  1.01177819  0.11711814 1.5980805  0.91497315 1.8450096
#> [27,]  0.02327165  0.47106102 -0.47152770 1.4493645  0.20780794 1.6884909
#> [28,]  0.31314882  0.72802646 -0.05633987 1.6499118  0.35315507 1.8420820
#> [29,]  0.09734702  0.66463256 -0.09379083 1.1482360  0.89948028 1.7616611
#> [30,]  0.32405914  0.50228546 -0.25486354 1.5711813  0.25134781 1.6974893
#> [31,]  0.19808514  0.33620935 -0.35630958 1.5664029 -0.15985588 1.6116143
#> [32,]  0.34034966  0.48885057 -0.35257596 1.3179657  0.27255870 1.5183490
#> [33,]  0.42350182  0.97071738  0.12607344 1.6712756  0.74409221 1.9346443
#> [34,] -0.17389009  0.32060482 -0.35852453 0.5015983  0.83221049 1.3082226
#> [35,]  0.33873364  0.80817167  0.13690079 1.2214919  0.63878711 1.5498277
#> [36,]  0.13756907 -0.19031874 -0.71007228 1.3969134 -0.91295629 1.0996148
#> [37,]  0.48402369  0.85486121 -0.01270895 1.9272212  0.33017882 1.7225698
#> [38,]  0.38318581  0.60193039 -0.27937923 1.8910701 -0.02210716 1.6809101
#> [39,]  0.59811383  1.00679922  0.07411703 1.9738435  0.55514835 1.8624162
#> [40,] -0.56445325 -0.09293985 -0.85540059 0.4532554  0.21549839 1.2498412
#>            [,7]        [,8]         [,9]       [,10]
#>  [1,] 0.8302522  0.21551612 -0.183303982  0.58478697
#>  [2,] 0.2455260 -0.09064427 -1.051331521 -0.05457380
#>  [3,] 0.4466382 -0.15811222 -1.011729203  0.02725909
#>  [4,] 1.1649683  0.72254547 -0.018484297  0.86032485
#>  [5,] 0.9410132  0.31195793 -0.398829930  0.57336389
#>  [6,] 1.2711004  0.72190472  0.187422989  0.93579595
#>  [7,] 0.8239441  0.57122494 -0.392461660  0.51201719
#>  [8,] 0.8880171  0.82779450 -0.639832378  0.43240490
#>  [9,] 1.2305954  0.49688176 -0.207403071  0.89073634
#> [10,] 1.1815724  0.65927110  0.133313110  0.82659546
#> [11,] 0.9316178  0.44515121 -0.851765897  0.23518065
#> [12,] 0.5532759  0.14438106 -0.735698887  0.34053591
#> [13,] 1.2690454  0.40851627 -0.278177578  1.01124997
#> [14,] 0.8323223  0.57105310 -0.216036385  0.37238696
#> [15,] 1.3688383  0.79378649  0.325058730  0.99325869
#> [16,] 1.5223137  1.19290582  0.201021178  1.00690109
#> [17,] 0.7537239  0.41993116 -0.420648061  0.35110277
#> [18,] 1.2121226  0.63862871  0.170655046  0.80739988
#> [19,] 0.9319203  0.10480309 -0.486584441  0.59386164
#> [20,] 0.9914977  0.45107195 -0.614803792  0.67784215
#> [21,] 1.4482637  1.10296740  0.061769432  0.87226293
#> [22,] 1.1526708  0.54305519 -0.245126562  0.68424193
#> [23,] 0.6475290 -0.11488125 -0.660305448  0.43231985
#> [24,] 0.8976817  0.27266283 -0.123510914  0.66296084
#> [25,] 1.2601621  0.60333965  0.090387431  1.05143009
#> [26,] 1.3118658  0.74167455  0.355525449  1.04363949
#> [27,] 1.1055934  0.49821095 -0.406086708  0.57130278
#> [28,] 1.2970601  0.92174249  0.099687161  0.88639244
#> [29,] 1.1920181  0.45153295 -0.032223497  0.92714493
#> [30,] 1.3472975  0.81610654 -0.091303601  0.91216460
#> [31,] 1.1095788  0.88718932 -0.199487441  0.69059454
#> [32,] 1.0362770  0.73071265 -0.161362217  0.70799714
#> [33,] 1.3961258  0.83961863  0.325655074  1.03275631
#> [34,] 0.7310502  0.11271529 -0.441561718  0.66313288
#> [35,] 1.0547103  0.70184473  0.280284709  0.76883808
#> [36,] 0.8698937  0.79125655 -0.644924712  0.18670255
#> [37,] 1.3776301  0.96396040  0.326829564  0.93153412
#> [38,] 1.2817088  1.01256370 -0.007893465  0.75857198
#> [39,] 1.4552564  1.03076137  0.398286401  1.06750605
#> [40,] 0.5936611 -0.08765819 -0.922162973  0.23608791
#> 
#> $pre[[3]]
#>             [,1]        [,2]
#>  [1,] -0.2400158 -0.02573848
#>  [2,] -0.1664407 -0.03403107
#>  [3,] -0.1876138 -0.03896351
#>  [4,] -0.2371421 -0.03362539
#>  [5,] -0.2201086 -0.03725126
#>  [6,] -0.2553227 -0.02914630
#>  [7,] -0.1947063 -0.03747201
#>  [8,] -0.1638437 -0.04302344
#>  [9,] -0.2348926 -0.04037149
#> [10,] -0.2471954 -0.02818096
#> [11,] -0.1692888 -0.04539932
#> [12,] -0.2039593 -0.03718509
#> [13,] -0.2478783 -0.03955847
#> [14,] -0.2002712 -0.03169176
#> [15,] -0.2596666 -0.02683860
#> [16,] -0.2261069 -0.03512135
#> [17,] -0.2066505 -0.03211593
#> [18,] -0.2508449 -0.02770298
#> [19,] -0.2398572 -0.03853022
#> [20,] -0.2037478 -0.04436594
#> [21,] -0.2129227 -0.03478836
#> [22,] -0.2291613 -0.03780972
#> [23,] -0.2329041 -0.03562601
#> [24,] -0.2427290 -0.02589678
#> [25,] -0.2567977 -0.03160528
#> [26,] -0.2677923 -0.02495942
#> [27,] -0.2111899 -0.04318227
#> [28,] -0.2369591 -0.03249388
#> [29,] -0.2558020 -0.03112785
#> [30,] -0.2225071 -0.03577296
#> [31,] -0.2022178 -0.03701117
#> [32,] -0.2149601 -0.03501505
#> [33,] -0.2608656 -0.02791718
#> [34,] -0.2395577 -0.03183362
#> [35,] -0.2561421 -0.02310240
#> [36,] -0.1500030 -0.04121256
#> [37,] -0.2409344 -0.02798516
#> [38,] -0.2116331 -0.03607628
#> [39,] -0.2510525 -0.02783231
#> [40,] -0.2000364 -0.04080649
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4466347  0.4466347
#>  [2,]  0.5330541 -0.5330541
#>  [3,]  0.5370943 -0.5370943
#>  [4,]  0.5507043 -0.5507043
#>  [5,] -0.4544126  0.4544126
#>  [6,]  0.5563043 -0.5563043
#>  [7,] -0.4607722  0.4607722
#>  [8,]  0.5301684 -0.5301684
#>  [9,]  0.5484775 -0.5484775
#> [10,]  0.5545358 -0.5545358
#> [11,] -0.4690672  0.4690672
#> [12,] -0.4584028  0.4584028
#> [13,] -0.4481076  0.4481076
#> [14,]  0.5420453 -0.5420453
#> [15,]  0.5579455 -0.5579455
#> [16,]  0.5476018 -0.5476018
#> [17,]  0.5435232 -0.5435232
#> [18,]  0.5555552 -0.5555552
#> [19,]  0.5501624 -0.5501624
#> [20,]  0.5397613 -0.5397613
#> [21,] -0.4555838  0.4555838
#> [22,]  0.5476925 -0.5476925
#> [23,] -0.4508398  0.4508398
#> [24,] -0.4460033  0.4460033
#> [25,] -0.4439386  0.4439386
#> [26,]  0.5604117 -0.5604117
#> [27,]  0.5419034 -0.5419034
#> [28,] -0.4490610  0.4490610
#> [29,] -0.4440666  0.4440666
#> [30,]  0.5465483 -0.5465483
#> [31,] -0.4587920  0.4587920
#> [32,] -0.4551347  0.4551347
#> [33,] -0.4420248  0.4420248
#> [34,] -0.4482549  0.4482549
#> [35,] -0.4420023  0.4420023
#> [36,]  0.5271708 -0.5271708
#> [37,]  0.5530370 -0.5530370
#> [38,]  0.5437768 -0.5437768
#> [39,]  0.5555745 -0.5555745
#> [40,] -0.4602764  0.4602764
#> 
#> $L
#> [1] 0.6948875 0.7753964 0.7952215 0.7023579 0.6978142 0.7056133
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536