Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12          V13         V14
#>  [1,] -0.2926400 -0.08433636  0.004410274 -0.01949898  0.002454787 -0.05786916
#>  [2,] -0.1972627 -0.22566734 -0.036044730 -0.10419153  0.008685969  0.06630855
#>  [3,] -0.2815648 -0.17431403 -0.140910648 -0.12851069 -0.146495140 -0.03850716
#>  [4,] -0.3468526 -0.08556243 -0.127328153 -0.16602895 -0.048080123 -0.08954365
#>  [5,] -0.2939131 -0.01946317  0.005372857 -0.09646593 -0.118130597 -0.03634976
#>  [6,] -0.3115327 -0.01762412 -0.034140045  0.02946293 -0.021929676  0.06915200
#>  [7,] -0.4176685 -0.23548203 -0.209295851 -0.19199703 -0.093886859 -0.06783906
#>  [8,] -0.2490301 -0.18470575 -0.133604311 -0.05888859 -0.025320611 -0.14157476
#>  [9,] -0.3236205 -0.11223140 -0.075745416 -0.15913568 -0.189883219 -0.18311569
#> [10,] -0.3075618 -0.07845101 -0.130867462 -0.05402063  0.030932963 -0.09441494
#>                 V15         V16         V17         V18         V19         V2
#>  [1,] -9.545196e-02  0.03853794  0.04203142  0.10264150  0.03262967 -0.2809614
#>  [2,] -2.638547e-02 -0.05373046  0.11107902  0.15621540  0.20095632 -0.2145864
#>  [3,]  4.686759e-02  0.03760996  0.07720419  0.05566439  0.19274240 -0.3217935
#>  [4,] -1.479178e-01  0.01741358 -0.11208761 -0.06822081  0.04225037 -0.3728757
#>  [5,] -1.406452e-03  0.02784892  0.01922707  0.11969759  0.03764649 -0.2935839
#>  [6,]  6.211389e-02 -0.03801121  0.06432853  0.06645941  0.08638673 -0.2515069
#>  [7,] -9.240529e-02 -0.08642279 -0.03271847 -0.01028584 -0.08715243 -0.2504788
#>  [8,]  2.678986e-05 -0.04009762  0.05773304  0.06211614  0.13120802 -0.3423359
#>  [9,] -1.810715e-02 -0.06707072 -0.16492232 -0.16169380 -0.14506321 -0.1984041
#> [10,] -2.898550e-02  0.03196046 -0.02063899  0.05756931  0.04813087 -0.2684222
#>               V20         V21         V22         V23        V24         V25
#>  [1,]  0.07259919  0.17709385  0.24197619  0.12582198 0.26091338  0.12784492
#>  [2,]  0.13481052  0.18757823  0.14653171  0.22377847 0.17699216  0.20608761
#>  [3,]  0.15305542  0.20862912  0.21869070  0.20927102 0.22579166  0.24987150
#>  [4,]  0.08639577  0.13114858  0.18346600  0.21088402 0.10791038  0.20082652
#>  [5,]  0.16113924  0.15771951  0.15990768  0.27688466 0.19702705  0.16137072
#>  [6,]  0.08198409  0.04590510  0.09121584  0.07739752 0.08060504  0.14526769
#>  [7,]  0.01713438  0.10965181 -0.02152535  0.10753931 0.05953710  0.12242284
#>  [8,]  0.13604350  0.06015166  0.16794606  0.24198066 0.21957394  0.19860315
#>  [9,] -0.07979301 -0.13887361 -0.11748920 -0.04483874 0.04520787 -0.09841467
#> [10,]  0.16053967  0.05419757  0.06331694  0.12507180 0.25271977  0.24616184
#>             V26         V27        V28          V29         V3         V30
#>  [1,] 0.1286523  0.17071580 0.20529988  0.034330739 -0.2669968 -0.10373177
#>  [2,] 0.1043821  0.11755028 0.05682362  0.123077282 -0.1774290  0.07556529
#>  [3,] 0.2113122  0.24789898 0.26201834  0.080516830 -0.3148102  0.04630794
#>  [4,] 0.1323335  0.16449284 0.01918511  0.130973607 -0.2962521 -0.04669786
#>  [5,] 0.1885023  0.28117855 0.16106138  0.200572056 -0.3272521  0.11604339
#>  [6,] 0.2395731  0.13302766 0.20630707  0.173135919 -0.2547048  0.07548575
#>  [7,] 0.1239218  0.15158575 0.07337111  0.070148435 -0.3424755  0.05920382
#>  [8,] 0.1479611  0.05886359 0.15089081 -0.064432479 -0.2393508  0.03650295
#>  [9,] 0.0845065 -0.07670374 0.02235284  0.005668987 -0.1874625  0.07300952
#> [10,] 0.1278779  0.22423886 0.12021000  0.131988324 -0.2416637  0.09498138
#>                V31          V32          V33          V34           V35
#>  [1,] -0.082721013 -0.015693384 -0.165838352 -0.090137050 -0.0947143114
#>  [2,] -0.046621031 -0.074547502  0.109088318 -0.012312850 -0.1103646377
#>  [3,]  0.107875488 -0.034058371  0.058614742 -0.098449333 -0.0008339034
#>  [4,]  0.035267555 -0.162340946 -0.012772047 -0.037897917  0.0409098754
#>  [5,] -0.047484201  0.022936451  0.032179931 -0.032928001  0.0747044245
#>  [6,]  0.055050636 -0.052302568 -0.071724069  0.009488221 -0.0442031900
#>  [7,]  0.005497903 -0.072266600 -0.044098491 -0.106319603 -0.0409101076
#>  [8,] -0.003655916 -0.008054386 -0.091848073 -0.045310108 -0.1302294480
#>  [9,]  0.071572819  0.003417611 -0.001550091  0.002347498 -0.0005955111
#> [10,] -0.009715079  0.089286532  0.036272544  0.142102323  0.0706672426
#>               V36          V37         V38          V39         V4         V40
#>  [1,] -0.23553011 -0.068529519 -0.02874235 -0.111147652 -0.3259080 -0.05527866
#>  [2,] -0.07476209 -0.038460561 -0.11069707 -0.113522862 -0.3061643 -0.12617777
#>  [3,]  0.02569045 -0.051400031 -0.02787428 -0.074045612 -0.2956645 -0.01200179
#>  [4,] -0.04819844 -0.061278238 -0.02853215 -0.092974498 -0.1520044 -0.06989828
#>  [5,] -0.02729054 -0.008579038  0.02931112 -0.042392081 -0.2949112 -0.05631017
#>  [6,] -0.01578193  0.030511058 -0.06563398 -0.130119732 -0.2609393 -0.12551034
#>  [7,] -0.09275660 -0.157148984 -0.13041807  0.001386469 -0.2035858 -0.04271325
#>  [8,] -0.15118730 -0.071747058 -0.19894642 -0.135561699 -0.3764782 -0.05176879
#>  [9,] -0.02595690 -0.116285478 -0.10903083 -0.181848776 -0.1989532 -0.05770367
#> [10,]  0.19821872  0.053296251 -0.03361856  0.082508536 -0.2793951  0.14951380
#>               V41          V42         V43         V44          V45         V46
#>  [1,] -0.08523937 -0.082126106 -0.16533324 -0.20693841 -0.219726992 -0.11563162
#>  [2,] -0.04491682 -0.142689885 -0.07145608 -0.03700113 -0.057633143 -0.18706311
#>  [3,] -0.01753713 -0.012980736 -0.10056188 -0.03638844 -0.103081893 -0.11354206
#>  [4,] -0.13303943 -0.158307412 -0.02148569 -0.12172710 -0.121413127 -0.12920596
#>  [5,]  0.04542508 -0.005944266 -0.01161661 -0.14898448  0.007448031 -0.06087796
#>  [6,] -0.14072238 -0.139188493  0.02695968 -0.16664484 -0.144658157 -0.16795541
#>  [7,] -0.05427308 -0.036925929 -0.15419604 -0.05318119 -0.164760088 -0.13232253
#>  [8,] -0.04065201 -0.088744736 -0.20242805 -0.07837613 -0.159566592 -0.26848528
#>  [9,] -0.06413043 -0.097928330 -0.06744237 -0.07559134 -0.158639346 -0.10886419
#> [10,] -0.05143640  0.081887622  0.01597508 -0.04497677 -0.096276973 -0.12358118
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.1083505 -0.1622699 -0.2951705 -0.2249228 -0.2607009 -0.3325143
#>  [2,] -0.2530265 -0.2790734 -0.2121418 -0.1520034 -0.2720230 -0.3105227
#>  [3,] -0.2358986 -0.2514361 -0.3412401 -0.1812817 -0.3212470 -0.4474745
#>  [4,] -0.1332538 -0.2779413 -0.1937708 -0.1587022 -0.3234461 -0.3477992
#>  [5,] -0.1580984 -0.2639071 -0.3224361 -0.2275523 -0.4062231 -0.2956843
#>  [6,] -0.1253146 -0.2039235 -0.1771014 -0.2130642 -0.3267757 -0.2198933
#>  [7,] -0.2511363 -0.2130993 -0.3056929 -0.3699816 -0.2603724 -0.3176973
#>  [8,] -0.2489586 -0.1730986 -0.3013631 -0.2298816 -0.2789551 -0.4123101
#>  [9,] -0.1213348 -0.2020469 -0.2340479 -0.2276501 -0.2514770 -0.3495740
#> [10,] -0.1110185 -0.2160126 -0.1514333 -0.1416885 -0.2537976 -0.2692174
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3983705 -0.2657882 -0.2947290 -0.2660751 -0.2666748 -0.2268914
#>  [2,] -0.4051822 -0.2766275 -0.2895437 -0.3488909 -0.4186286 -0.3063630
#>  [3,] -0.3403324 -0.4256472 -0.4837232 -0.4135407 -0.3385246 -0.3410898
#>  [4,] -0.3061334 -0.2957062 -0.3580290 -0.3303160 -0.3428244 -0.3229648
#>  [5,] -0.2411177 -0.3233436 -0.3080746 -0.4185028 -0.4252454 -0.3279080
#>  [6,] -0.2782270 -0.2404993 -0.3318397 -0.2180790 -0.2410068 -0.3193641
#>  [7,] -0.3105469 -0.4690058 -0.3386731 -0.3598262 -0.2602125 -0.3805916
#>  [8,] -0.3037216 -0.3597841 -0.3137574 -0.2874636 -0.4445763 -0.4774439
#>  [9,] -0.1955315 -0.3835580 -0.2594829 -0.3764158 -0.2710216 -0.2320667
#> [10,] -0.3953255 -0.2689319 -0.2957689 -0.3788796 -0.3208038 -0.3172721
#>              V58        V59          V6        V60         V7          V8
#>  [1,] -0.3656726 -0.3654356 -0.12518065 -0.3038183 -0.2618908 -0.16031657
#>  [2,] -0.2439600 -0.3607792 -0.13238086 -0.2834735 -0.1567172 -0.14744982
#>  [3,] -0.4415085 -0.3722922 -0.21049902 -0.3607701 -0.1664463 -0.15483800
#>  [4,] -0.4000790 -0.2170151 -0.19576893 -0.3519292 -0.1773574 -0.21552365
#>  [5,] -0.2698393 -0.2859569 -0.25248347 -0.2898006 -0.1176933 -0.17554607
#>  [6,] -0.2468060 -0.3432636 -0.24146489 -0.2696220 -0.1149289 -0.14271131
#>  [7,] -0.3156410 -0.3556782 -0.27537263 -0.3316702 -0.2446800 -0.18202953
#>  [8,] -0.4007762 -0.3767984 -0.23046794 -0.4013683 -0.1242981 -0.18551606
#>  [9,] -0.2584555 -0.3279509 -0.20395623 -0.2601156 -0.2331878 -0.26659461
#> [10,] -0.3423575 -0.3374818 -0.09207024 -0.3603890 -0.2217593 -0.07058356
#>                 V9
#>  [1,] -0.103970968
#>  [2,] -0.232568344
#>  [3,] -0.200214650
#>  [4,] -0.215276942
#>  [5,] -0.110658248
#>  [6,] -0.003824972
#>  [7,] -0.092714652
#>  [8,] -0.228207645
#>  [9,] -0.269299332
#> [10,] -0.100888190
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]         [,4]        [,5]       [,6]
#> [1,] -0.05089587  0.03332462  0.005564096 0.0394711351 -0.02616577 0.03122980
#> [2,] -0.05725222 -0.04152594 -0.054899604 0.0004399118 -0.08686695 0.04850003
#>             [,7]        [,8]       [,9]       [,10]
#> [1,] -0.02083796 -0.09020684 -0.1752802 -0.12037784
#> [2,]  0.12135913 -0.08414834  0.1628273  0.08122173
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.557737e-05  4.623984e-05  2.486130e-05  4.830094e-05  5.541637e-05
#>  [2,] -1.168620e-04 -4.505906e-04 -6.321333e-04 -4.929700e-04 -4.143015e-04
#>  [3,] -1.003941e-04 -3.977654e-04 -5.105276e-04 -3.584055e-04 -3.033789e-04
#>  [4,] -1.077412e-04 -4.492480e-04 -7.039056e-04 -6.785215e-04 -6.212176e-04
#>  [5,] -8.843100e-05 -3.222814e-04 -4.428100e-04 -3.337809e-04 -2.860571e-04
#>  [6,] -1.139303e-05  2.178416e-05  7.939475e-06  2.498213e-05  3.049056e-05
#>  [7,]  1.071146e-04  4.900882e-04  7.871624e-04  7.994549e-04  7.286320e-04
#>  [8,] -2.701155e-05 -4.529921e-05 -1.064469e-04 -9.239896e-05 -7.938931e-05
#>  [9,]  1.889956e-04  8.472323e-04  1.494720e-03  1.451377e-03  1.173029e-03
#> [10,]  1.552956e-04  7.396205e-04  1.004363e-03  9.116274e-04  8.377807e-04
#>                 V14           V15           V16           V17           V18
#>  [1,]  6.298665e-05  7.590030e-05  1.160285e-04  1.594679e-04  1.576647e-04
#>  [2,] -3.438514e-04 -2.152019e-04 -1.840931e-04  6.761522e-05  5.524024e-05
#>  [3,] -3.178524e-04 -2.686669e-04 -2.561900e-04 -3.353778e-05 -2.410608e-05
#>  [4,] -4.758135e-04 -2.988243e-04 -2.770884e-04 -8.902893e-05 -1.574882e-04
#>  [5,] -2.755122e-04 -2.192804e-04 -2.007181e-04 -1.896738e-05 -2.129606e-05
#>  [6,]  3.158930e-05  4.107676e-05  6.636732e-05  1.044798e-04  1.057514e-04
#>  [7,]  5.339674e-04  3.149801e-04  3.005248e-04  1.512264e-04  2.512668e-04
#>  [8,] -4.344958e-05  1.290452e-06  2.750613e-05  8.435576e-05  7.399068e-05
#>  [9,]  5.855921e-04 -3.690900e-05 -1.193766e-04 -4.548609e-04 -3.474253e-04
#> [10,]  7.583845e-04  6.368245e-04  6.617948e-04  3.946425e-04  4.601871e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  1.243745e-04  4.197516e-07  1.293880e-04  1.581662e-04  3.052121e-04
#>  [2,] -4.087336e-04 -9.846911e-05 -7.617439e-04 -7.151082e-04 -1.573993e-04
#>  [3,] -3.834728e-04 -8.625183e-05 -7.147535e-04 -6.864775e-04 -2.279063e-04
#>  [4,] -6.394008e-04 -9.883604e-05 -8.911411e-04 -8.409829e-04 -4.329709e-04
#>  [5,] -3.411560e-04 -7.269204e-05 -6.031501e-04 -5.617722e-04 -1.774818e-04
#>  [6,]  7.099359e-05 -2.024874e-06  6.640604e-05  8.996645e-05  1.942831e-04
#>  [7,]  7.431811e-04  1.011996e-04  9.799051e-04  9.444992e-04  5.984815e-04
#>  [8,] -3.546950e-05 -1.673898e-05 -7.022303e-05 -4.252884e-05  1.143294e-04
#>  [9,]  5.932018e-04  1.734866e-04  9.464582e-04  7.493450e-04 -8.420191e-05
#> [10,]  1.070461e-03  1.521448e-04  1.524160e-03  1.513039e-03  1.007059e-03
#>                 V23           V24           V25           V26           V27
#>  [1,]  2.812736e-04  2.820943e-04  2.340462e-04  2.597692e-04  3.269846e-04
#>  [2,]  1.566966e-04  3.531040e-04  2.796007e-04  3.255218e-04  5.640411e-04
#>  [3,]  4.941952e-05  1.924715e-04  1.328520e-04  1.518188e-04  3.557538e-04
#>  [4,] -1.641406e-04  8.984566e-05  1.268496e-04  2.151925e-04  3.910726e-04
#>  [5,]  5.135197e-05  2.056502e-04  1.794522e-04  2.160933e-04  3.850890e-04
#>  [6,]  1.991648e-04  2.096342e-04  1.856091e-04  2.048503e-04  2.531344e-04
#>  [7,]  3.598081e-04  1.038431e-04  6.208382e-05 -3.906246e-05 -1.707589e-04
#>  [8,]  1.573469e-04  2.113844e-04  1.949006e-04  2.288814e-04  3.006664e-04
#>  [9,] -7.100867e-04 -1.345938e-03 -1.350567e-03 -1.703997e-03 -1.904832e-03
#> [10,]  5.988887e-04  2.932598e-04  2.167822e-04  1.806772e-04 -2.925663e-06
#>                 V28           V29            V3           V30           V31
#>  [1,]  2.642074e-04  2.706442e-04  1.317719e-06  0.0002764797  0.0002505357
#>  [2,]  3.844021e-04  5.532727e-04 -4.912690e-05  0.0006886784  0.0010738669
#>  [3,]  2.433391e-04  3.587517e-04 -4.662411e-05  0.0003993673  0.0007182818
#>  [4,]  2.411462e-04  4.409218e-04 -5.614862e-05  0.0006882791  0.0010576043
#>  [5,]  2.844359e-04  3.856846e-04 -4.077611e-05  0.0004462938  0.0007177102
#>  [6,]  2.110445e-04  2.139358e-04  1.458102e-06  0.0002125373  0.0001978256
#>  [7,] -3.143711e-05 -2.677427e-04  6.390103e-05 -0.0006132372 -0.0010236780
#>  [8,]  2.418714e-04  2.836280e-04 -7.910635e-06  0.0003310728  0.0003825224
#>  [9,] -1.579114e-03 -2.238703e-03  8.412330e-05 -0.0030126523 -0.0036504107
#> [10,]  1.566746e-04 -4.087312e-05  9.398388e-05 -0.0002623389 -0.0008411764
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0002088208  0.0001585757  0.0001442043  0.0001231246  0.0001153145
#>  [2,]  0.0007939590  0.0006818679  0.0007853439  0.0008878836  0.0009174086
#>  [3,]  0.0005299577  0.0004170267  0.0004500531  0.0005037733  0.0005462125
#>  [4,]  0.0007628599  0.0007013924  0.0008797409  0.0009891463  0.0009881667
#>  [5,]  0.0005374392  0.0004330559  0.0004728207  0.0005244352  0.0005556420
#>  [6,]  0.0001607942  0.0001359997  0.0001335802  0.0001134890  0.0001044773
#>  [7,] -0.0007499398 -0.0007161598 -0.0009191030 -0.0010658363 -0.0010695611
#>  [8,]  0.0002962007  0.0002571999  0.0002876011  0.0002998504  0.0002933328
#>  [9,] -0.0028042386 -0.0028402340 -0.0034995065 -0.0037796527 -0.0035972341
#> [10,] -0.0005379256 -0.0004092330 -0.0005427358 -0.0006994675 -0.0007866310
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0001211675  1.187216e-04  1.356717e-04 -1.200570e-06  0.0001884304
#>  [2,]  0.0006730429  1.907486e-04  1.085522e-04 -1.396099e-04  0.0005618087
#>  [3,]  0.0003914321  7.518124e-05  5.137605e-05 -1.170149e-04  0.0003850105
#>  [4,]  0.0007132738  1.328866e-04 -2.181898e-05 -1.483609e-04  0.0004177411
#>  [5,]  0.0004085885  1.101871e-04  8.562902e-05 -1.026380e-04  0.0003843162
#>  [6,]  0.0001027983  9.091823e-05  9.846502e-05 -2.571996e-06  0.0001416193
#>  [7,] -0.0007734538 -1.267042e-04  6.189790e-05  1.559330e-04 -0.0003737160
#>  [8,]  0.0002358275  1.201267e-04  9.630656e-05 -2.676954e-05  0.0002096667
#>  [9,] -0.0026935653 -1.182323e-03 -7.557525e-04  2.802165e-04 -0.0017215174
#> [10,] -0.0004925861  1.464781e-04  2.726603e-04  2.173780e-04 -0.0002479198
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0001856903  1.063968e-04  5.977340e-05  7.285528e-05  1.117023e-04
#>  [2,]  0.0004904131  1.007243e-04 -1.574290e-04 -1.595502e-04 -3.040750e-04
#>  [3,]  0.0003522920  7.637231e-05 -1.174849e-04 -1.322748e-04 -2.459571e-04
#>  [4,]  0.0003255342 -5.121403e-05 -2.829245e-04 -2.210640e-04 -3.363906e-04
#>  [5,]  0.0003418629  7.979073e-05 -9.867310e-05 -9.742338e-05 -1.840985e-04
#>  [6,]  0.0001359442  8.247893e-05  4.473552e-05  5.003632e-05  7.285583e-05
#>  [7,] -0.0002731527  1.283392e-04  3.688944e-04  2.997814e-04  4.141786e-04
#>  [8,]  0.0001836419  6.696722e-05 -3.580726e-06  1.415892e-05  9.567827e-06
#>  [9,] -0.0014104497 -4.290047e-04  2.243639e-04  2.031131e-04  4.977386e-04
#> [10,] -0.0001828468  2.251620e-04  4.794132e-04  4.123933e-04  6.100339e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.204411e-04  7.841190e-05  2.509192e-05  1.826226e-05  3.442967e-05
#>  [2,] -1.611623e-04 -3.990226e-05 -1.520104e-04 -9.639801e-05 -9.735062e-05
#>  [3,] -1.280406e-04 -1.311490e-05 -1.006580e-04 -6.289866e-05 -8.532320e-05
#>  [4,] -2.138103e-04 -1.153900e-04 -1.925084e-04 -1.271689e-04 -1.419474e-04
#>  [5,] -8.757035e-05 -2.340423e-06 -8.259528e-05 -5.239954e-05 -7.365863e-05
#>  [6,]  7.749021e-05  5.315201e-05  1.678609e-05  1.143247e-05  1.700339e-05
#>  [7,]  2.719730e-04  1.671176e-04  2.240012e-04  1.475976e-04  1.583439e-04
#>  [8,]  3.215622e-05  2.333138e-05 -2.186188e-05 -1.393916e-05 -5.963880e-06
#>  [9,]  2.627850e-04  1.938334e-04  4.196397e-04  2.743746e-04  2.366269e-04
#> [10,]  4.368637e-04  2.150627e-04  2.675866e-04  1.723896e-04  2.286078e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.454223e-05  1.188985e-05  2.318316e-06  1.629644e-06  2.832775e-07
#>  [2,]  6.703012e-06 -4.893633e-06 -2.270316e-05 -1.321259e-05 -1.952636e-05
#>  [3,]  3.535211e-06 -3.349693e-06 -2.034759e-05 -1.211053e-05 -1.806228e-05
#>  [4,] -1.624383e-06 -1.716729e-05 -2.594476e-05 -1.510637e-05 -2.350076e-05
#>  [5,]  5.814813e-06 -2.758933e-06 -1.755468e-05 -1.029012e-05 -1.638379e-05
#>  [6,]  9.940024e-06  7.828772e-06  1.325638e-06  1.209033e-06 -2.634779e-07
#>  [7,]  5.057289e-06  2.301998e-05  2.855211e-05  1.721043e-05  2.376817e-05
#>  [8,]  8.366001e-06  3.695495e-06 -2.911665e-06 -1.209551e-06 -3.734156e-06
#>  [9,] -4.291071e-05  1.487722e-05  3.743596e-05  1.386123e-05  2.460950e-05
#> [10,]  2.409080e-05  3.640049e-05  4.360855e-05  2.788393e-05  3.836586e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  2.377898e-06 -1.161912e-06 -4.536854e-07 -3.018642e-06 -1.377782e-06
#>  [2,] -6.355154e-06 -1.398312e-05  8.667842e-08 -2.089247e-05 -1.745589e-05
#>  [3,] -7.047023e-06 -1.180878e-05 -2.643172e-06 -1.709664e-05 -1.445224e-05
#>  [4,] -7.908522e-06 -1.473148e-05  8.709762e-07 -2.300844e-05 -2.070416e-05
#>  [5,] -5.516583e-06 -1.001174e-05 -1.994313e-06 -1.551639e-05 -1.338521e-05
#>  [6,]  1.708179e-06 -8.265648e-07  1.567351e-07 -2.233673e-06 -1.330199e-06
#>  [7,]  8.190152e-06  1.463664e-05 -1.079216e-06  2.323618e-05  2.067329e-05
#>  [8,]  3.202753e-07 -3.075170e-06  5.624747e-07 -5.631356e-06 -4.386974e-06
#>  [9,] -4.296795e-06  1.712084e-05 -2.231709e-05  3.579272e-05  2.874855e-05
#> [10,]  1.852372e-05  2.198726e-05  5.050853e-06  3.121310e-05  2.880552e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.814419e-05  2.422750e-06  3.908981e-05  2.210971e-05  2.469273e-06
#>  [2,] -2.917461e-05 -4.584478e-06  2.419124e-06 -2.035089e-04 -5.243093e-04
#>  [3,] -3.181263e-05 -4.663955e-06 -1.398923e-05 -1.917031e-04 -4.635338e-04
#>  [4,] -8.488202e-05 -6.657521e-06 -4.660002e-05 -2.048246e-04 -4.603292e-04
#>  [5,] -2.491696e-05 -3.751477e-06 -4.076140e-06 -1.498259e-04 -3.758622e-04
#>  [6,]  2.420859e-05  1.312797e-06  2.850634e-05  1.510267e-05 -6.804547e-06
#>  [7,]  1.190206e-04  6.768216e-06  7.772924e-05  2.307525e-04  4.829816e-04
#>  [8,]  1.097307e-05  4.376157e-07  2.099481e-05 -1.492455e-05 -7.189893e-05
#>  [9,]  9.886441e-05 -1.055902e-06 -5.638625e-05  2.838690e-04  8.806092e-04
#> [10,]  1.743998e-04  1.376503e-05  1.471341e-04  3.801530e-04  7.299994e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.01239847  0.01793388  0.02322289  0.02221446  0.02195144  0.01863623
#> [2,] -0.01239847 -0.01793388 -0.02322289 -0.02221446 -0.02195144 -0.01863623
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.02632294  0.01929463  0.02912842  0.02583342
#> [2,] -0.02632294 -0.01929463 -0.02912842 -0.02583342
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4622681 0.5949424 0.7497957 0.3640361 0.7208202 0.3770589 0.4155946
#>  [8] 0.4949936 0.1747376 0.6804367
#> 
#> $B[[2]]
#> [1]  0.045528189 -0.004392958
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  4.198491e-04  8.595054e-04  4.633950e-04  6.343566e-04  5.162949e-04
#>  [6]  3.190055e-04 -4.881627e-04  4.192884e-04 -3.764083e-03  3.158099e-05
#> 
#> $vB[[2]]
#> [1]  0.02620788 -0.02620788
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#>  [2,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [3,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#>  [4,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#>  [5,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#>  [6,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#>  [7,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#>  [8,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#>  [9,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [10,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [11,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [12,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [13,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [14,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [15,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [16,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [17,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [18,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [19,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [20,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [21,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [22,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [23,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [24,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [25,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [26,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [27,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [28,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [29,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [30,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [31,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [32,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [33,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [34,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [35,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [36,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [37,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [38,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [39,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [40,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#>  [2,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [3,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#>  [4,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#>  [5,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#>  [6,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#>  [7,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#>  [8,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#>  [9,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [10,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [11,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [12,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [13,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [14,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [15,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [16,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [17,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [18,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [19,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [20,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [21,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [22,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [23,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [24,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [25,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [26,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [27,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [28,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [29,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [30,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [31,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [32,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [33,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [34,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [35,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [36,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [37,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [38,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [39,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [40,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#>  [2,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [3,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#>  [4,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#>  [5,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#>  [6,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#>  [7,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#>  [8,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#>  [9,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [10,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [11,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [12,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [13,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [14,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [15,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [16,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [17,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [18,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [19,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [20,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [21,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [22,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [23,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [24,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [25,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [26,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [27,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [28,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [29,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [30,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [31,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [32,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [33,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [34,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [35,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [36,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [37,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [38,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [39,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [40,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#>  [2,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [3,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#>  [4,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#>  [5,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#>  [6,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#>  [7,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#>  [8,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#>  [9,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [10,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [11,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [12,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [13,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [14,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [15,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [16,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [17,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [18,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [19,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [20,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [21,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [22,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [23,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [24,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [25,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [26,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [27,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [28,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [29,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [30,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [31,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [32,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [33,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [34,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [35,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [36,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [37,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [38,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [39,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [40,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#>  [2,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [3,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#>  [4,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#>  [5,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#>  [6,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#>  [7,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#>  [8,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#>  [9,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [10,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [11,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [12,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [13,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [14,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [15,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [16,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [17,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [18,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [19,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [20,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [21,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [22,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [23,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [24,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [25,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [26,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [27,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [28,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [29,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [30,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [31,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [32,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [33,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [34,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [35,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [36,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [37,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [38,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [39,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [40,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#>  [2,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [3,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#>  [4,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#>  [5,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#>  [6,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#>  [7,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#>  [8,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#>  [9,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [10,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [11,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [12,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [13,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [14,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [15,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [16,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [17,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [18,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [19,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [20,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [21,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [22,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [23,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [24,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [25,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [26,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [27,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [28,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [29,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [30,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [31,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [32,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [33,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [34,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [35,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [36,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [37,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [38,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [39,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [40,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5466717 0.7059025 0.8078376 0.5188472 0.8207223 0.6066922 0.3729471
#>  [2,] 0.3725327 0.5092539 0.6958508 0.3833127 0.7532778 0.5386967 0.2863712
#>  [3,] 0.7230814 0.7822634 0.8747904 0.6876917 0.8714223 0.7265043 0.6072095
#>  [4,] 0.6723613 0.7990505 0.8752205 0.6126293 0.8611432 0.6809832 0.4734465
#>  [5,] 0.6801225 0.8004770 0.8720029 0.5993836 0.8622619 0.6975244 0.4644345
#>  [6,] 0.5735004 0.7349675 0.8360571 0.5410385 0.8498150 0.6290848 0.4256791
#>  [7,] 0.6547117 0.7390793 0.8408600 0.6010092 0.8520275 0.7292091 0.5451619
#>  [8,] 0.6426654 0.7845808 0.8806694 0.6792466 0.8843390 0.7168249 0.5488559
#>  [9,] 0.3734300 0.5270948 0.6849084 0.3939902 0.7447866 0.5459540 0.2742381
#> [10,] 0.4400846 0.5486540 0.6818903 0.4222718 0.7325234 0.5400957 0.3307751
#> [11,] 0.3862688 0.5474775 0.6910848 0.3863522 0.7539571 0.5522445 0.2956220
#> [12,] 0.6291598 0.7654790 0.8664393 0.6423718 0.8669228 0.7409536 0.5667964
#> [13,] 0.7796633 0.8271002 0.9008954 0.7013987 0.8960862 0.7838373 0.5906606
#> [14,] 0.7769035 0.8292953 0.9034362 0.7016401 0.9007622 0.7850418 0.5794483
#> [15,] 0.7426108 0.8076779 0.8926960 0.6899908 0.8887382 0.7799647 0.6093552
#> [16,] 0.6791221 0.8070365 0.8671791 0.6190570 0.8572135 0.7044658 0.4473520
#> [17,] 0.4920910 0.5947380 0.7571204 0.4834146 0.7757719 0.5277202 0.4233898
#> [18,] 0.5990204 0.7007790 0.7970339 0.5549207 0.8009374 0.6411679 0.4270472
#> [19,] 0.6991349 0.7477346 0.8512399 0.6702170 0.8468289 0.7204326 0.6310576
#> [20,] 0.6122293 0.7101861 0.8189409 0.5650184 0.8366913 0.7045072 0.5074626
#> [21,] 0.7205851 0.8307571 0.8886343 0.6155243 0.8846215 0.7762021 0.4999938
#> [22,] 0.5569173 0.7320812 0.8312437 0.5963324 0.8387504 0.6980110 0.5216816
#> [23,] 0.5135045 0.6537292 0.8139252 0.5289219 0.8369266 0.6764518 0.4537425
#> [24,] 0.5823513 0.6826913 0.8103225 0.5127579 0.8039226 0.6998371 0.4452989
#> [25,] 0.4162959 0.5541627 0.7075390 0.4718284 0.7404215 0.5485668 0.4206953
#> [26,] 0.6736881 0.7822253 0.8330471 0.5298679 0.8240709 0.6951490 0.3975010
#> [27,] 0.7944646 0.8699562 0.9203776 0.7311886 0.9026099 0.7810227 0.6264284
#> [28,] 0.3124012 0.4453394 0.6051285 0.3296591 0.6742391 0.4362685 0.2869907
#> [29,] 0.6505160 0.7120378 0.8135655 0.5204941 0.8288931 0.7052927 0.4598203
#> [30,] 0.5049809 0.6420389 0.7963289 0.5226539 0.8286596 0.6292037 0.4157930
#> [31,] 0.7057876 0.7910468 0.8641614 0.6368529 0.8556496 0.7111835 0.5004822
#> [32,] 0.7775847 0.8419998 0.9212079 0.6931482 0.9121628 0.7898030 0.5922190
#> [33,] 0.6549611 0.7499964 0.8441934 0.6254463 0.8489986 0.7253309 0.5804491
#> [34,] 0.7231689 0.8156492 0.8717074 0.6004655 0.8601250 0.7365391 0.4661555
#> [35,] 0.8032405 0.8681331 0.9257420 0.7199235 0.9048830 0.7850721 0.6071579
#> [36,] 0.6402488 0.7418365 0.8464468 0.5530945 0.8418442 0.6878468 0.4244868
#> [37,] 0.6252332 0.7480926 0.7992428 0.5024013 0.7984589 0.6322401 0.3888464
#> [38,] 0.6688328 0.7847346 0.8832729 0.6683559 0.8835252 0.7432585 0.5810197
#> [39,] 0.6548090 0.7797665 0.8744991 0.5872990 0.8637609 0.6865121 0.4353373
#> [40,] 0.7486962 0.8299825 0.8988759 0.6461129 0.8854093 0.7539009 0.5084029
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5380168 0.1953768 0.8435471
#>  [2,] 0.3139347 0.1702344 0.7886053
#>  [3,] 0.7155778 0.3665029 0.8619164
#>  [4,] 0.6749000 0.2176415 0.8680307
#>  [5,] 0.6702768 0.1951964 0.8540508
#>  [6,] 0.5800777 0.2154130 0.8725935
#>  [7,] 0.6308178 0.3380718 0.8446794
#>  [8,] 0.6728366 0.3537907 0.8941219
#>  [9,] 0.3267551 0.1814041 0.7906187
#> [10,] 0.3664518 0.1938960 0.7727759
#> [11,] 0.3271313 0.1683090 0.7850332
#> [12,] 0.6437872 0.3813815 0.8711419
#> [13,] 0.7478864 0.2874621 0.8629261
#> [14,] 0.7460519 0.2699182 0.8752350
#> [15,] 0.7384876 0.3631689 0.8721905
#> [16,] 0.6712105 0.1996464 0.8513230
#> [17,] 0.4782750 0.2690845 0.8139410
#> [18,] 0.5878166 0.2557703 0.8148136
#> [19,] 0.6929031 0.4190629 0.8396868
#> [20,] 0.5828419 0.3219668 0.8253806
#> [21,] 0.7035989 0.2235419 0.8524196
#> [22,] 0.5996270 0.3910091 0.8530256
#> [23,] 0.5053726 0.2974189 0.8652928
#> [24,] 0.5492720 0.2541558 0.8208541
#> [25,] 0.4134799 0.3456673 0.8228937
#> [26,] 0.6300158 0.1638733 0.8032267
#> [27,] 0.8045133 0.3043462 0.8772359
#> [28,] 0.2887230 0.2181539 0.7761984
#> [29,] 0.5913947 0.2504433 0.8144936
#> [30,] 0.4946039 0.2827570 0.8464978
#> [31,] 0.6855203 0.2447090 0.8465566
#> [32,] 0.7593660 0.2655990 0.8923402
#> [33,] 0.6517326 0.3823867 0.8420764
#> [34,] 0.6991403 0.1824830 0.8305679
#> [35,] 0.8020341 0.2668484 0.8758730
#> [36,] 0.6068531 0.1952667 0.8294438
#> [37,] 0.6124217 0.1971098 0.7892417
#> [38,] 0.6749451 0.3602466 0.8834140
#> [39,] 0.6427236 0.1899670 0.8794700
#> [40,] 0.7249151 0.2068263 0.8592995
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5567867 0.4432133
#>  [2,] 0.5461114 0.4538886
#>  [3,] 0.5511170 0.4488830
#>  [4,] 0.5618458 0.4381542
#>  [5,] 0.5639205 0.4360795
#>  [6,] 0.5565641 0.4434359
#>  [7,] 0.5495312 0.4504688
#>  [8,] 0.5514625 0.4485375
#>  [9,] 0.5458778 0.4541222
#> [10,] 0.5460040 0.4539960
#> [11,] 0.5474549 0.4525451
#> [12,] 0.5476748 0.4523252
#> [13,] 0.5607175 0.4392825
#> [14,] 0.5619415 0.4380585
#> [15,] 0.5531830 0.4468170
#> [16,] 0.5644505 0.4355495
#> [17,] 0.5440884 0.4559116
#> [18,] 0.5532603 0.4467397
#> [19,] 0.5448036 0.4551964
#> [20,] 0.5488499 0.4511501
#> [21,] 0.5643784 0.4356216
#> [22,] 0.5440223 0.4559777
#> [23,] 0.5463141 0.4536859
#> [24,] 0.5516590 0.4483410
#> [25,] 0.5338905 0.4661095
#> [26,] 0.5651103 0.4348897
#> [27,] 0.5612519 0.4387481
#> [28,] 0.5347196 0.4652804
#> [29,] 0.5544148 0.4455852
#> [30,] 0.5473537 0.4526463
#> [31,] 0.5605143 0.4394857
#> [32,] 0.5623945 0.4376055
#> [33,] 0.5467391 0.4532609
#> [34,] 0.5667464 0.4332536
#> [35,] 0.5642949 0.4357051
#> [36,] 0.5610964 0.4389036
#> [37,] 0.5596630 0.4403370
#> [38,] 0.5508767 0.4491233
#> [39,] 0.5628073 0.4371927
#> [40,] 0.5662610 0.4337390
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]      [,3]         [,4]      [,5]       [,6]
#>  [1,]  0.18723189  0.87556569 1.4360202  0.075424607 1.5212489  0.4334291
#>  [2,] -0.52136672  0.03701991 0.8276170 -0.475511025 1.1161708  0.1550969
#>  [3,]  0.95979818  1.27890515 1.9439951  0.789349608 1.9135933  0.9769585
#>  [4,]  0.71888418  1.38037036 1.9479275  0.458378070 1.8248174  0.7582937
#>  [5,]  0.75433477  1.38927849 1.9187855  0.402897422 1.8342047  0.8355369
#>  [6,]  0.29614710  1.01997360 1.6291789  0.164524010 1.7331506  0.5282927
#>  [7,]  0.63981508  1.04118890 1.6646407  0.409671835 1.7505927  0.9906136
#>  [8,]  0.58695146  1.29256344 1.9987843  0.750311681 2.0341772  0.9287663
#>  [9,] -0.51753009  0.10848539 0.7764215 -0.430570193 1.0709976  0.1843361
#> [10,] -0.24081881  0.19523388 0.7624727 -0.313454460 1.0074630  0.1607281
#> [11,] -0.46302346  0.19048384 0.8051957 -0.462671773 1.1198295  0.2097434
#> [12,]  0.52861392  1.18295689 1.8698360  0.585673393 1.8740203  1.0509309
#> [13,]  1.26370523  1.56521388 2.2072131  0.853967434 2.1544752  1.2881701
#> [14,]  1.24771172  1.58064114 2.2360021  0.855120106 2.2057225  1.2952934
#> [15,]  1.05958276  1.43499204 2.1185799  0.800076180 2.0779171  1.2654607
#> [16,]  0.74974004  1.43086810 1.8762440  0.485547785 1.7923361  0.8686551
#> [17,] -0.03163884  0.38358730 1.1369565 -0.066366033 1.2411945  0.1109948
#> [18,]  0.40138503  0.85101000 1.3678581  0.220572839 1.3921637  0.5804366
#> [19,]  0.84318172  1.08656618 1.7443593  0.709166525 1.7099434  0.9466086
#> [20,]  0.45669254  0.89628820 1.5091885  0.261554427 1.6338131  0.8688543
#> [21,]  0.94736552  1.59100229 2.0768669  0.470594033 2.0369420  1.2436695
#> [22,]  0.22866027  1.00520744 1.5944674  0.390206460 1.6489593  0.8378441
#> [23,]  0.05403122  0.63547191 1.4757196  0.115816859 1.6355361  0.7375133
#> [24,]  0.33243307  0.76616788 1.4521073  0.051042876 1.4109936  0.8465222
#> [25,] -0.33799786  0.21750405 0.8834613 -0.112805824 1.0481607  0.1948817
#> [26,]  0.72491355  1.27868180 1.6073788  0.119614046 1.5441753  0.8243031
#> [27,]  1.35205006  1.90057170 2.4474880  1.000661357 2.2265657  1.2716363
#> [28,] -0.78891704 -0.21951963 0.4268805 -0.709727229 0.7274213 -0.2563203
#> [29,]  0.62130804  0.90530190 1.4733466  0.082022225 1.5778027  0.8726302
#> [30,]  0.01992422  0.58422475 1.3635056  0.090677677 1.5761574  0.5288022
#> [31,]  0.87501254  1.33124686 1.8502921  0.561730589 1.7796175  0.9011389
#> [32,]  1.25164620  1.67318357 2.4588732  0.814878927 2.3403340  1.3237385
#> [33,]  0.64091885  1.09859307 1.6897660  0.512730488 1.7267684  0.9710611
#> [34,]  0.96023533  1.48714360 1.9161406  0.407404999 1.8163288  1.0280572
#> [35,]  1.40667207  1.88455154 2.5230497  0.944082254 2.2526985  1.2954728
#> [36,]  0.57644430  1.05553589 1.7070000  0.213181682 1.6720142  0.7900719
#> [37,]  0.51182085  1.08846524 1.3815686  0.009605144 1.3766903  0.5418387
#> [38,]  0.70291065  1.29347373 2.0237954  0.700758480 2.0262446  1.0629742
#> [39,]  0.64024551  1.26430643 1.9413386  0.352810793 1.8468844  0.7838631
#> [40,]  1.09167072  1.58550308 2.1847962  0.601996368 2.0446838  1.1195267
#>                [,7]        [,8]       [,9]    [,10]
#>  [1,] -5.195944e-01  0.15236134 -1.4154444 1.684861
#>  [2,] -9.130739e-01 -0.78178759 -1.5839671 1.316539
#>  [3,]  4.355976e-01  0.92263065 -0.5472490 1.831299
#>  [4,] -1.063142e-01  0.73043152 -1.2794636 1.883658
#>  [5,] -1.425028e-01  0.70943714 -1.4165921 1.766732
#>  [6,] -2.995024e-01  0.32309239 -1.2926004 1.924087
#>  [7,]  1.811412e-01  0.53572688 -0.6718989 1.693466
#>  [8,]  1.960491e-01  0.72104292 -0.6024179 2.133554
#>  [9,] -9.732252e-01 -0.72289823 -1.5068635 1.328659
#> [10,] -7.046816e-01 -0.54746900 -1.4248906 1.224052
#> [11,] -8.682335e-01 -0.72118838 -1.5976596 1.295242
#> [12,]  2.687923e-01  0.59184003 -0.4836887 1.911093
#> [13,]  3.666970e-01  1.08737112 -0.9077420 1.839809
#> [14,]  3.205091e-01  1.07766535 -0.9950375 1.948061
#> [15,]  4.446027e-01  1.03812261 -0.5616364 1.920467
#> [16,] -2.113754e-01  0.71366499 -1.3885061 1.745016
#> [17,] -3.088731e-01 -0.08695470 -0.9992724 1.475824
#> [18,] -2.939089e-01  0.35494678 -1.0680702 1.481597
#> [19,]  5.367566e-01  0.81372673 -0.3266214 1.655899
#> [20,]  2.985274e-02  0.33445064 -0.7447478 1.553236
#> [21,] -2.468654e-05  0.86449473 -1.2451436 1.753706
#> [22,]  8.678068e-02  0.40391128 -0.4430726 1.758531
#> [23,] -1.855605e-01  0.02149108 -0.8596191 1.859964
#> [24,] -2.196835e-01  0.19772980 -1.0765692 1.522145
#> [25,] -3.199199e-01 -0.34959809 -0.6381393 1.536077
#> [26,] -4.158887e-01  0.53228477 -1.6296865 1.406585
#> [27,]  5.169249e-01  1.41474501 -0.8266863 1.966512
#> [28,] -9.100448e-01 -0.90159416 -1.2764572 1.243648
#> [29,] -1.610663e-01  0.36973385 -1.0962494 1.479477
#> [30,] -3.400678e-01 -0.02158523 -0.9308269 1.707393
#> [31,]  1.928723e-03  0.77925843 -1.1270332 1.707845
#> [32,]  3.731462e-01  1.14920669 -1.0170677 2.114871
#> [33,]  3.246173e-01  0.62666391 -0.4794300 1.673759
#> [34,] -1.355853e-01  0.84320750 -1.4996149 1.589657
#> [35,]  4.353810e-01  1.39905638 -1.0106719 1.953916
#> [36,] -3.043811e-01  0.43410322 -1.4161447 1.581690
#> [37,] -4.521638e-01  0.45750338 -1.4044568 1.320361
#> [38,]  3.269609e-01  0.73063722 -0.5742939 2.025164
#> [39,] -2.601076e-01  0.58720500 -1.4502246 1.987421
#> [40,]  3.361471e-02  0.96897492 -1.3441629 1.809484
#> 
#> $pre[[3]]
#>               [,1]       [,2]
#>  [1,]  0.024515595 -0.2036155
#>  [2,]  0.030733195 -0.1542382
#>  [3,]  0.001688533 -0.2034962
#>  [4,]  0.020986237 -0.2276702
#>  [5,]  0.025551537 -0.2315370
#>  [6,]  0.019779952 -0.2074490
#>  [7,]  0.007454951 -0.1913219
#>  [8,]  0.005538187 -0.2010433
#>  [9,]  0.029555972 -0.1544728
#> [10,]  0.026868622 -0.1576692
#> [11,]  0.032620743 -0.1577721
#> [12,]  0.003969859 -0.1873104
#> [13,]  0.015235240 -0.2288393
#> [14,]  0.017051305 -0.2319939
#> [15,]  0.003392669 -0.2101469
#> [16,]  0.026816022 -0.2324284
#> [17,]  0.010064109 -0.1667488
#> [18,]  0.016609737 -0.1972428
#> [19,] -0.004955820 -0.1846521
#> [20,]  0.010964801 -0.1850600
#> [21,]  0.024693047 -0.2342580
#> [22,]  0.002353033 -0.1741933
#> [23,]  0.013113636 -0.1726755
#> [24,]  0.019517291 -0.1878587
#> [25,]  0.002489550 -0.1332806
#> [26,]  0.031367567 -0.2305610
#> [27,]  0.011452848 -0.2347915
#> [28,]  0.015036770 -0.1240655
#> [29,]  0.017371339 -0.2011534
#> [30,]  0.014103433 -0.1758809
#> [31,]  0.019463636 -0.2237860
#> [32,]  0.015993425 -0.2348922
#> [33,]  0.001907524 -0.1855964
#> [34,]  0.028725733 -0.2398628
#> [35,]  0.016352759 -0.2422584
#> [36,]  0.026687739 -0.2189250
#> [37,]  0.023451654 -0.2163426
#> [38,]  0.005359143 -0.1988546
#> [39,]  0.024390712 -0.2281726
#> [40,]  0.024964542 -0.2416475
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5567867  0.5567867
#>  [2,]  0.4538886 -0.4538886
#>  [3,] -0.5511170  0.5511170
#>  [4,]  0.4381542 -0.4381542
#>  [5,]  0.4360795 -0.4360795
#>  [6,]  0.4434359 -0.4434359
#>  [7,]  0.4504688 -0.4504688
#>  [8,] -0.5514625  0.5514625
#>  [9,]  0.4541222 -0.4541222
#> [10,]  0.4539960 -0.4539960
#> [11,]  0.4525451 -0.4525451
#> [12,] -0.5476748  0.5476748
#> [13,]  0.4392825 -0.4392825
#> [14,]  0.4380585 -0.4380585
#> [15,]  0.4468170 -0.4468170
#> [16,]  0.4355495 -0.4355495
#> [17,] -0.5440884  0.5440884
#> [18,]  0.4467397 -0.4467397
#> [19,] -0.5448036  0.5448036
#> [20,]  0.4511501 -0.4511501
#> [21,] -0.5643784  0.5643784
#> [22,] -0.5440223  0.5440223
#> [23,]  0.4536859 -0.4536859
#> [24,] -0.5516590  0.5516590
#> [25,] -0.5338905  0.5338905
#> [26,]  0.4348897 -0.4348897
#> [27,] -0.5612519  0.5612519
#> [28,] -0.5347196  0.5347196
#> [29,] -0.5544148  0.5544148
#> [30,] -0.5473537  0.5473537
#> [31,] -0.5605143  0.5605143
#> [32,]  0.4376055 -0.4376055
#> [33,] -0.5467391  0.5467391
#> [34,]  0.4332536 -0.4332536
#> [35,] -0.5642949  0.5642949
#> [36,] -0.5610964  0.5610964
#> [37,]  0.4403370 -0.4403370
#> [38,] -0.5508767  0.5508767
#> [39,] -0.5628073  0.5628073
#> [40,]  0.4337390 -0.4337390
#> 
#> $L
#> [1] 0.6908927 0.6909235 0.6897589 0.6918307 0.6951460 0.6945021
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174