Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13         V14
#>  [1,] -0.4145116 -0.12612858 -0.080712286 -0.062199856 -0.16467252  0.01748932
#>  [2,] -0.2163538 -0.16969911 -0.047870098 -0.094536333 -0.01237569 -0.04254797
#>  [3,] -0.3510494 -0.06660378 -0.005923958 -0.025950054 -0.14202977 -0.10582722
#>  [4,] -0.2327003 -0.16094964 -0.146836890 -0.168527593 -0.03732048 -0.14735861
#>  [5,] -0.2739508 -0.04746509 -0.102475821 -0.043208675  0.05440280 -0.01546524
#>  [6,] -0.3195877 -0.02145317 -0.150069859 -0.094039343 -0.17755377 -0.02597908
#>  [7,] -0.2758795 -0.11328390 -0.181307199 -0.139144508 -0.18268273 -0.19118881
#>  [8,] -0.2466153 -0.05309718 -0.178497884  0.004421009  0.03593011  0.04609375
#>  [9,] -0.2373213 -0.12263935  0.012974982  0.038400782  0.02124407 -0.08092239
#> [10,] -0.3053901 -0.21990301 -0.093369074 -0.031601281 -0.12114278 -0.10973817
#>                 V15          V16         V17         V18         V19         V2
#>  [1,] -0.0329794697 -0.028366881  0.18445563  0.07785471  0.13830875 -0.4056787
#>  [2,] -0.0285265058  0.044243109  0.09067829  0.06122760  0.12943077 -0.3235905
#>  [3,] -0.1135320201  0.008479629 -0.09219990 -0.04652481  0.07343967 -0.2240437
#>  [4,] -0.0008034574 -0.085204273 -0.12239706  0.06802518 -0.08712672 -0.2479398
#>  [5,]  0.0294534756  0.178166110  0.08236910  0.13830788  0.23692834 -0.3172116
#>  [6,] -0.0133411651 -0.165034874  0.01148166 -0.10808936 -0.11379472 -0.2145333
#>  [7,] -0.0144342645 -0.056632578 -0.04577954 -0.13097664 -0.18890554 -0.1545788
#>  [8,]  0.0699758639  0.083795340  0.01746016  0.10906487  0.09341057 -0.2824261
#>  [9,] -0.1194208811  0.052374865 -0.08632340 -0.05803720 -0.01315367 -0.3072525
#> [10,] -0.0577402947 -0.171795231 -0.06594259  0.04815730  0.07431364 -0.2910958
#>                V20         V21         V22        V23         V24         V25
#>  [1,]  0.206804693  0.14340211  0.25500042 0.17474648  0.28958322  0.23961432
#>  [2,]  0.075520941  0.21232974  0.21370685 0.24042758  0.20593275  0.24779434
#>  [3,] -0.041679529  0.12073321  0.02715101 0.08869994  0.14792812  0.06919692
#>  [4,]  0.024488870  0.09521630  0.03667823 0.10336062  0.06489465  0.16593965
#>  [5,]  0.182156173  0.25147158  0.17860494 0.15276256  0.18229409  0.19727717
#>  [6,] -0.025438223  0.06753548  0.15400728 0.21919982  0.21878063  0.25293622
#>  [7,] -0.019075979 -0.17677035 -0.03539925 0.02745898 -0.06459481 -0.10586904
#>  [8,] -0.001778974  0.01474643  0.04680505 0.23460569  0.09604799  0.11812072
#>  [9,]  0.161134409  0.11702295  0.17216827 0.06300086  0.08253772 -0.02286788
#> [10,]  0.168105287  0.03508809  0.12285342 0.10321930  0.06222678  0.02919958
#>               V26           V27         V28          V29         V3
#>  [1,]  0.18307686  0.2342593158  0.25583634  0.148923602 -0.3057486
#>  [2,]  0.14564424  0.1469886926  0.03057550  0.008357209 -0.2735368
#>  [3,]  0.06689311  0.0363087947  0.15004188  0.150035209 -0.3386190
#>  [4,]  0.24362222  0.1693679025  0.06780932  0.105136477 -0.3081692
#>  [5,]  0.18196856  0.2224726248  0.22207154  0.073665636 -0.3054252
#>  [6,]  0.24188368  0.1512650664  0.11506679  0.114117605 -0.3613475
#>  [7,] -0.08831793  0.0724197620 -0.04268694 -0.144313967 -0.2065706
#>  [8,]  0.19532562  0.1720224608  0.14168785  0.023081554 -0.2844027
#>  [9,]  0.03512316 -0.0003161739  0.01840895  0.166115345 -0.2929962
#> [10,]  0.14347049  0.1900553254  0.18108347  0.021125479 -0.2187419
#>                 V30         V31         V32         V33         V34
#>  [1,]  0.1418109309  0.04638991 -0.02855494  0.01850035  0.01539842
#>  [2,] -0.0363698700 -0.02459576 -0.00362584  0.09778703 -0.06107323
#>  [3,] -0.0688866657 -0.04824516  0.07385831  0.03158753 -0.07520377
#>  [4,]  0.1030224363  0.05041697 -0.06601178 -0.07096078  0.09331135
#>  [5,]  0.1921964173  0.18384070  0.10726606 -0.05210225  0.02491708
#>  [6,]  0.0743013886  0.05186462  0.09718271  0.13315416  0.10438264
#>  [7,]  0.0094297113 -0.10600737 -0.00379499 -0.15981194  0.01689337
#>  [8,]  0.0093181620 -0.04827400 -0.02961634 -0.01725682 -0.09333270
#>  [9,]  0.0489346474  0.09421503 -0.05084875 -0.03365770 -0.05311256
#> [10,] -0.0008770988 -0.02171636 -0.08043412 -0.19512997 -0.16749309
#>                 V35           V36         V37          V38           V39
#>  [1,]  0.0753360752  0.0681203777  0.05695769 -0.097227532 -0.0294931910
#>  [2,] -0.1067459209 -0.1073414884  0.04875921 -0.098897839  0.0550614142
#>  [3,]  0.0533687339 -0.1279523413 -0.16443854 -0.053046330 -0.1295604026
#>  [4,]  0.0005679817  0.0493999254  0.05957796  0.040920125 -0.0667126841
#>  [5,]  0.0135176143 -0.0658818216 -0.05563740 -0.059859045  0.0318478877
#>  [6,] -0.0611217865 -0.0333448814 -0.05915311 -0.039992480  0.0171384114
#>  [7,] -0.0991544184 -0.0482932022 -0.07500857 -0.109973028 -0.0000666864
#>  [8,] -0.0807939735 -0.1120064286 -0.23070516 -0.163267085 -0.1235072477
#>  [9,]  0.0071347086  0.0000861972 -0.04748872  0.005553918 -0.0602158311
#> [10,] -0.0579534638 -0.0738109524 -0.11246574 -0.080436063 -0.1974856826
#>               V4           V40          V41         V42          V43
#>  [1,] -0.3021083  0.0131065167 -0.102713100 -0.12308413 -0.134554720
#>  [2,] -0.1652711  0.0006383737 -0.078234989 -0.07672590  0.003999944
#>  [3,] -0.2870753 -0.1354515265 -0.173825839 -0.17949708 -0.117876186
#>  [4,] -0.3647773 -0.1581067743 -0.006410972 -0.03396455 -0.135016354
#>  [5,] -0.2332648 -0.0471623829 -0.011250221 -0.03617679  0.029535300
#>  [6,] -0.3414772 -0.0368893679 -0.111255481 -0.06796785 -0.104067097
#>  [7,] -0.1876492 -0.0566197852  0.043647455 -0.01784456 -0.069878069
#>  [8,] -0.3479744 -0.1089696521 -0.165149933 -0.03578887 -0.104905744
#>  [9,] -0.2231024 -0.0587614304 -0.003252921 -0.15052504 -0.006831461
#> [10,] -0.2444695 -0.2143420500 -0.193675408 -0.05423983 -0.130977961
#>               V44         V45         V46        V47        V48        V49
#>  [1,] -0.03010769 -0.08948617 -0.15164097 -0.2104074 -0.1582200 -0.3475830
#>  [2,] -0.14893568 -0.22670014 -0.17353207 -0.1067261 -0.2415607 -0.1687357
#>  [3,] -0.09836774 -0.11621467 -0.05708218 -0.1109133 -0.2863127 -0.1822863
#>  [4,] -0.18349026 -0.12761273 -0.17461155 -0.2353909 -0.1563586 -0.1829315
#>  [5,] -0.03241954 -0.11465459 -0.15188339 -0.2221482 -0.2212541 -0.2998007
#>  [6,] -0.03786867 -0.03453313 -0.20892921 -0.1319419 -0.2878539 -0.3334640
#>  [7,] -0.10071056 -0.24028589 -0.22012450 -0.1301907 -0.2191026 -0.2406803
#>  [8,] -0.17324106 -0.19306194 -0.11126653 -0.2330479 -0.1779642 -0.2188899
#>  [9,] -0.18246730 -0.04195375 -0.08135359 -0.2250646 -0.2664240 -0.2490563
#> [10,] -0.23087698 -0.16557021 -0.10780075 -0.1931133 -0.3425417 -0.3558608
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.1895904 -0.3812748 -0.4087875 -0.3289846 -0.3164717 -0.3491587
#>  [2,] -0.2003277 -0.3315468 -0.2419184 -0.2467170 -0.3680447 -0.2813922
#>  [3,] -0.3123180 -0.2293510 -0.2329874 -0.3998086 -0.2680576 -0.3881148
#>  [4,] -0.2248327 -0.2512640 -0.2892293 -0.3248330 -0.3922897 -0.3548964
#>  [5,] -0.1632967 -0.3463437 -0.3866899 -0.3431607 -0.3828151 -0.4046864
#>  [6,] -0.2569084 -0.2282629 -0.3883855 -0.3164604 -0.3492841 -0.3192306
#>  [7,] -0.2183025 -0.1762891 -0.3583654 -0.3310450 -0.2170759 -0.2964417
#>  [8,] -0.2865657 -0.3629709 -0.3578943 -0.2481860 -0.3509683 -0.2782560
#>  [9,] -0.3119176 -0.3340075 -0.2697708 -0.3389768 -0.2360483 -0.2463744
#> [10,] -0.2291072 -0.3381670 -0.3264233 -0.2599671 -0.2597328 -0.2519261
#>              V55        V56        V57        V58        V59          V6
#>  [1,] -0.2930972 -0.4222758 -0.3278795 -0.3274644 -0.3887299 -0.14932030
#>  [2,] -0.3071988 -0.2348376 -0.3504570 -0.2746269 -0.3392829 -0.12132062
#>  [3,] -0.3762116 -0.2721495 -0.3601677 -0.3270434 -0.3288430 -0.20272148
#>  [4,] -0.2903717 -0.4010916 -0.2715253 -0.3689248 -0.3999126 -0.30429538
#>  [5,] -0.3533673 -0.3417416 -0.4099836 -0.2560144 -0.2822672 -0.27203431
#>  [6,] -0.3544002 -0.2349148 -0.2499799 -0.3031345 -0.2888546 -0.11642805
#>  [7,] -0.2895196 -0.1964356 -0.3513695 -0.1933180 -0.3445308 -0.13157200
#>  [8,] -0.2293404 -0.3444802 -0.3721483 -0.2848998 -0.3531284 -0.21967657
#>  [9,] -0.3532434 -0.3688874 -0.2408455 -0.4104080 -0.3611959 -0.09998364
#> [10,] -0.3745579 -0.3702000 -0.3721764 -0.4306267 -0.3457851 -0.15369921
#>              V60          V7         V8          V9
#>  [1,] -0.3915578 -0.25613289 -0.2192113 -0.12216817
#>  [2,] -0.3930507 -0.15298047 -0.2153785 -0.22458945
#>  [3,] -0.3206603 -0.11557857 -0.1062913 -0.17413898
#>  [4,] -0.2677472 -0.22873489 -0.1738937 -0.11131303
#>  [5,] -0.3659906 -0.07678632 -0.1852827 -0.03202004
#>  [6,] -0.3377656 -0.11635224 -0.1376060 -0.17297202
#>  [7,] -0.3434691 -0.17724774 -0.2555966 -0.13317501
#>  [8,] -0.3928511 -0.25747227 -0.2325922 -0.21133264
#>  [9,] -0.3529073 -0.07999459 -0.1655937 -0.03664977
#> [10,] -0.2620275 -0.26819506 -0.1061507 -0.13557743
#> 
#> $W[[2]]
#>            [,1]       [,2]       [,3]        [,4]       [,5]        [,6]
#> [1,] 0.09977471 0.11020291 0.09955268 -0.05813405 0.02381847 -0.05803351
#> [2,] 0.01386197 0.02680223 0.07606126  0.04375517 0.03229851 -0.07048285
#>            [,7]         [,8]        [,9]       [,10]
#> [1,] -0.1010765  0.012012532  0.08050855  0.07772951
#> [2,]  0.1208739 -0.002851598 -0.05929116 -0.06523279
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -2.605693e-05 -6.978897e-05 -0.0001538018 -7.549989e-05 -1.527144e-05
#>  [2,] -2.640060e-05 -7.499707e-05 -0.0001811350 -9.982089e-05 -1.339383e-05
#>  [3,]  2.198093e-05  2.179706e-04  0.0001708312  1.791450e-04  2.471726e-04
#>  [4,]  2.165545e-04  1.392274e-03  0.0016575523  1.583024e-03  1.543662e-03
#>  [5,]  8.047275e-05  4.784803e-04  0.0005557026  5.175699e-04  5.121611e-04
#>  [6,]  4.048782e-05  3.527019e-04  0.0003311974  3.446729e-04  4.027503e-04
#>  [7,]  1.974007e-04  1.226135e-03  0.0016810220  1.600751e-03  1.345722e-03
#>  [8,]  6.956667e-05  4.701910e-04  0.0005262761  5.212756e-04  5.556961e-04
#>  [9,] -1.773504e-04 -9.358239e-04 -0.0013505516 -1.245822e-03 -1.020564e-03
#> [10,] -1.991736e-04 -1.128919e-03 -0.0016380319 -1.595030e-03 -1.337619e-03
#>                 V14           V15           V16          V17           V18
#>  [1,]  6.672019e-05  0.0001655890  3.106342e-04 3.892683e-04  3.815026e-04
#>  [2,]  1.244798e-04  0.0002749404  4.831625e-04 6.042412e-04  5.762582e-04
#>  [3,]  3.840029e-04  0.0004762815  6.544934e-04 7.469243e-04  7.591206e-04
#>  [4,]  1.495079e-03  0.0013290989  1.384416e-03 1.417003e-03  1.660931e-03
#>  [5,]  5.377292e-04  0.0005303140  5.669896e-04 5.638802e-04  6.307719e-04
#>  [6,]  5.272371e-04  0.0006053877  7.801170e-04 8.515939e-04  8.843494e-04
#>  [7,]  1.006398e-03  0.0007350761  4.950765e-04 4.561130e-04  7.361591e-04
#>  [8,]  6.201082e-04  0.0006502288  7.811229e-04 8.631171e-04  9.249418e-04
#>  [9,] -6.132448e-04 -0.0002106139  1.657188e-04 3.384662e-04  9.704037e-05
#> [10,] -8.325282e-04 -0.0004022442 -1.605425e-05 9.945336e-05 -2.013070e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0001534993 -1.926156e-05 -1.035749e-04 -8.639227e-05  0.0001625280
#>  [2,]  0.0002837256 -1.913797e-05 -2.022889e-05  1.711654e-05  0.0003338122
#>  [3,]  0.0006019207  3.375223e-05  5.404541e-04  6.087727e-04  0.0007540947
#>  [4,]  0.0020743112  2.427208e-04  2.810813e-03  2.934608e-03  0.0024717879
#>  [5,]  0.0007154596  9.326671e-05  9.853381e-04  1.050668e-03  0.0008517669
#>  [6,]  0.0007403463  5.766829e-05  7.199236e-04  7.768067e-04  0.0008589525
#>  [7,]  0.0013568507  2.120485e-04  2.160340e-03  2.225780e-03  0.0015930805
#>  [8,]  0.0009566775  8.310182e-05  1.098928e-03  1.179969e-03  0.0011735153
#>  [9,] -0.0007819781 -1.732153e-04 -1.736537e-03 -1.716303e-03 -0.0008928623
#> [10,] -0.0011294662 -2.021472e-04 -2.077544e-03 -2.090977e-03 -0.0013077642
#>                 V23           V24           V25          V26           V27
#>  [1,]  0.0003964351  4.341739e-04  0.0004278272 0.0004824448  5.781645e-04
#>  [2,]  0.0006123649  6.696847e-04  0.0006733274 0.0007340541  8.352683e-04
#>  [3,]  0.0008959908  8.997974e-04  0.0009016771 0.0009882683  1.128166e-03
#>  [4,]  0.0021505442  1.917755e-03  0.0017858986 0.0018977788  2.263730e-03
#>  [5,]  0.0007127308  6.379255e-04  0.0005839682 0.0006270333  7.448153e-04
#>  [6,]  0.0009629845  9.367106e-04  0.0009201979 0.0010166148  1.206253e-03
#>  [7,]  0.0009656599  5.800682e-04  0.0002553108 0.0001139655  5.169203e-04
#>  [8,]  0.0011715673  1.117258e-03  0.0010795836 0.0011297133  1.280784e-03
#>  [9,] -0.0001844414  8.796197e-05  0.0002486138 0.0003664316  3.151279e-04
#> [10,] -0.0006067684 -2.769455e-04 -0.0000328448 0.0001185326 -6.813076e-05
#>                 V28          V29            V3          V30          V31
#>  [1,]  5.631886e-04 5.862626e-04 -2.323996e-06 5.463768e-04 0.0006867354
#>  [2,]  7.938211e-04 8.314139e-04  5.468426e-06 8.033980e-04 0.0009534719
#>  [3,]  1.133965e-03 1.020772e-03  3.846637e-05 1.032602e-03 0.0010741795
#>  [4,]  2.596525e-03 1.962142e-03  1.826876e-04 2.038818e-03 0.0016271779
#>  [5,]  8.923979e-04 7.295517e-04  7.040964e-05 7.771960e-04 0.0006095432
#>  [6,]  1.266724e-03 1.104545e-03  5.500919e-05 1.131597e-03 0.0011595786
#>  [7,]  1.077022e-03 5.160579e-04  1.277173e-04 6.373542e-04 0.0001544122
#>  [8,]  1.360451e-03 1.149252e-03  7.218417e-05 1.148681e-03 0.0010388604
#>  [9,] -4.595017e-05 3.418353e-04 -9.577414e-05 3.243159e-04 0.0009096543
#> [10,] -4.928229e-04 4.909935e-05 -1.217181e-04 8.328818e-05 0.0006630238
#>                V32           V33           V34           V35           V36
#>  [1,] 0.0004770709  0.0004820620  0.0005618938  0.0007520141  0.0009631180
#>  [2,] 0.0006731116  0.0006861324  0.0008027803  0.0010742644  0.0012940367
#>  [3,] 0.0008512837  0.0007941402  0.0008721661  0.0010265120  0.0011271939
#>  [4,] 0.0016137138  0.0013063253  0.0012092548  0.0008308098  0.0005255753
#>  [5,] 0.0005751180  0.0004794851  0.0004854217  0.0003932783  0.0002874871
#>  [6,] 0.0009396489  0.0008463378  0.0009153713  0.0010265124  0.0011304517
#>  [7,] 0.0004222993 -0.0000173920 -0.0003798788 -0.0010602977 -0.0014962393
#>  [8,] 0.0008906616  0.0007671948  0.0007721916  0.0007749585  0.0007371044
#>  [9,] 0.0003977246  0.0006065626  0.0009274123  0.0017148817  0.0023198702
#> [10,] 0.0001562286  0.0004678219  0.0008726181  0.0017409736  0.0023180885
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0008480405  0.0004996574  3.231008e-04 -2.930145e-05  0.0005183351
#>  [2,]  0.0011127511  0.0006618887  3.987291e-04 -2.824670e-05  0.0006339583
#>  [3,]  0.0009651126  0.0006610537  5.188870e-04  3.439923e-05  0.0006250942
#>  [4,]  0.0004720535  0.0008529617  1.329050e-03  2.947378e-04  0.0009350559
#>  [5,]  0.0002648355  0.0003710964  5.087173e-04  1.126938e-04  0.0003676822
#>  [6,]  0.0009845732  0.0007295338  6.586115e-04  6.039282e-05  0.0007340799
#>  [7,] -0.0011687746 -0.0003337063  3.890482e-04  2.640276e-04 -0.0001881644
#>  [8,]  0.0006457521  0.0005793617  6.215324e-04  1.005873e-04  0.0005764071
#>  [9,]  0.0019481602  0.0007777005 -6.785922e-05 -2.328280e-04  0.0006349382
#> [10,]  0.0018945775  0.0006866917 -2.629843e-04 -2.665922e-04  0.0004541666
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0004375069  0.0002097719  8.750919e-05  6.835454e-06 -0.0001158163
#>  [2,]  0.0005417925  0.0002488505  1.014088e-04  1.350659e-05 -0.0001334160
#>  [3,]  0.0004789946  0.0003126495  2.503633e-04  1.720817e-04  0.0001047390
#>  [4,]  0.0005486273  0.0008871917  1.092794e-03  9.171373e-04  0.0010586568
#>  [5,]  0.0002162852  0.0003104984  3.753918e-04  3.048815e-04  0.0003848378
#>  [6,]  0.0005332945  0.0003893722  3.448147e-04  2.466471e-04  0.0001949203
#>  [7,] -0.0003947990  0.0003466183  7.190582e-04  6.661972e-04  0.0009115820
#>  [8,]  0.0004299333  0.0004325894  4.338158e-04  3.478798e-04  0.0003440118
#>  [9,]  0.0006421058 -0.0001533070 -5.280232e-04 -5.585359e-04 -0.0008796818
#> [10,]  0.0005043688 -0.0003627239 -7.419566e-04 -7.109160e-04 -0.0009896977
#>                 V46           V47           V48           V49            V5
#>  [1,] -9.947885e-05 -1.664279e-06 -3.384916e-06 -6.362167e-06 -1.702125e-05
#>  [2,] -1.157431e-04 -1.820468e-07 -4.255963e-06 -9.292668e-06 -1.728253e-05
#>  [3,]  7.048069e-05  9.761710e-05  9.134739e-05  4.164279e-05  4.421643e-05
#>  [4,]  8.007004e-04  5.116207e-04  4.878359e-04  2.696285e-04  3.504708e-04
#>  [5,]  3.086184e-04  1.925298e-04  1.771257e-04  1.003967e-04  1.243324e-04
#>  [6,]  1.385202e-04  1.441240e-04  1.388359e-04  6.985831e-05  7.533232e-05
#>  [7,]  7.117905e-04  4.553668e-04  4.140142e-04  2.242934e-04  2.903426e-04
#>  [8,]  2.615707e-04  2.048583e-04  1.825637e-04  9.767134e-05  1.289395e-04
#>  [9,] -6.966029e-04 -3.263777e-04 -2.983954e-04 -1.833456e-04 -2.718127e-04
#> [10,] -7.687320e-04 -4.170026e-04 -3.762084e-04 -2.271561e-04 -3.263404e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.008876e-05 -6.931850e-06 -1.093893e-06  8.093362e-06  1.219851e-06
#>  [2,]  1.408490e-05 -6.688903e-06  8.059519e-07  1.140494e-05  4.268168e-06
#>  [3,]  3.128849e-05  8.696770e-06  1.486564e-05  1.831026e-05  1.321184e-05
#>  [4,]  1.003579e-04  7.176459e-05  6.844590e-05  4.659066e-05  5.244125e-05
#>  [5,]  4.187167e-05  2.478726e-05  2.415906e-05  1.751643e-05  1.991434e-05
#>  [6,]  4.108805e-05  1.373773e-05  2.010276e-05  2.205058e-05  1.701478e-05
#>  [7,]  5.163386e-05  6.174544e-05  5.024615e-05  1.217430e-05  2.634695e-05
#>  [8,]  4.364472e-05  2.477585e-05  2.582003e-05  2.130010e-05  2.108219e-05
#>  [9,] -2.809001e-05 -5.667705e-05 -3.591780e-05 -2.914679e-06 -2.232818e-05
#> [10,] -3.643651e-05 -6.760687e-05 -4.444847e-05 -5.275259e-06 -2.531060e-05
#>                 V55           V56           V57           V58           V59
#>  [1,] -4.385227e-07 -9.646296e-07  4.394082e-06 -2.166765e-06  7.294150e-07
#>  [2,]  9.471973e-07 -2.618480e-07  7.423424e-06 -1.669451e-06  2.673331e-06
#>  [3,]  1.166146e-05  9.425765e-06  1.247742e-05  8.368640e-06  7.968304e-06
#>  [4,]  5.685106e-05  5.163487e-05  3.444354e-05  5.440083e-05  3.644387e-05
#>  [5,]  2.121205e-05  1.914399e-05  1.299478e-05  1.924361e-05  1.319602e-05
#>  [6,]  1.716844e-05  1.425760e-05  1.527972e-05  1.306803e-05  1.074954e-05
#>  [7,]  3.209282e-05  3.568123e-05  1.328447e-05  4.182643e-05  1.898740e-05
#>  [8,]  2.051781e-05  1.847152e-05  1.570320e-05  1.917105e-05  1.470735e-05
#>  [9,] -2.885151e-05 -3.084846e-05 -4.963750e-06 -3.648202e-05 -1.866301e-05
#> [10,] -3.256301e-05 -3.467297e-05 -6.511275e-06 -4.340881e-05 -2.261187e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  6.349286e-05  2.844849e-06  6.255624e-05 -1.535352e-06 -9.642862e-05
#>  [2,]  8.654589e-05  4.526987e-06  9.279472e-05  4.967618e-06 -1.020328e-04
#>  [3,]  1.336683e-04  9.194533e-06  1.579748e-04  1.760256e-04  1.644617e-04
#>  [4,]  3.867984e-04  3.132651e-05  4.385333e-04  8.532359e-04  1.177427e-03
#>  [5,]  1.397956e-04  1.147050e-05  1.583226e-04  3.226973e-04  4.198109e-04
#>  [6,]  1.677484e-04  1.170794e-05  1.930727e-04  2.619026e-04  2.720110e-04
#>  [7,]  2.357956e-04  1.957779e-05  2.583876e-04  7.171979e-04  1.058635e-03
#>  [8,]  1.860734e-04  1.352271e-05  2.132001e-04  3.091028e-04  3.872386e-04
#>  [9,] -8.699414e-05 -1.063872e-05 -8.259390e-05 -4.609720e-04 -8.456763e-04
#> [10,] -1.696126e-04 -1.503722e-05 -1.654976e-04 -5.778868e-04 -9.780098e-04
#> 
#> $vW[[2]]
#>            [,1]       [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.0306172 -0.0279714 -0.01587856 -0.01087207 -0.03678058 -0.01895962
#> [2,]  0.0306172  0.0279714  0.01587856  0.01087207  0.03678058  0.01895962
#>              [,7]        [,8]        [,9]       [,10]
#> [1,]  0.006323476 -0.02055974 -0.02064091 -0.01266135
#> [2,] -0.006323476  0.02055974  0.02064091  0.01266135
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.68907714 0.65994006 0.47323643 0.53454577 0.75931646 0.48331758
#>  [7] 0.01245752 0.34994909 0.49837585 0.33498678
#> 
#> $B[[2]]
#> [1]  0.03475050 -0.01040128
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0008877187 0.0012687596 0.0015770814 0.0030800573 0.0011316686
#>  [6] 0.0017485112 0.0003902321 0.0017453547 0.0006519938 0.0002289091
#> 
#> $vB[[2]]
#> [1] -0.03862509  0.03862509
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#>  [2,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#>  [3,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#>  [4,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#>  [5,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [6,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#>  [7,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#>  [8,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [9,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [10,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [11,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [12,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [13,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [14,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [15,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [16,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [17,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [18,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [19,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [20,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [21,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [22,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [23,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [24,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [25,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [26,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [27,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [28,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [29,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [30,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [31,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [32,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [33,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [34,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [35,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [36,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [37,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [38,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [39,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [40,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#>  [2,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#>  [3,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#>  [4,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#>  [5,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [6,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#>  [7,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#>  [8,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [9,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [10,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [11,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [12,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [13,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [14,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [15,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [16,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [17,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [18,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [19,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [20,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [21,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [22,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [23,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [24,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [25,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [26,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [27,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [28,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [29,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [30,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [31,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [32,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [33,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [34,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [35,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [36,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [37,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [38,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [39,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [40,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#>  [2,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#>  [3,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#>  [4,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#>  [5,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [6,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#>  [7,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#>  [8,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [9,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [10,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [11,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [12,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [13,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [14,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [15,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [16,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [17,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [18,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [19,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [20,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [21,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [22,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [23,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [24,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [25,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [26,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [27,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [28,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [29,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [30,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [31,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [32,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [33,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [34,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [35,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [36,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [37,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [38,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [39,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [40,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#>  [2,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#>  [3,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#>  [4,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#>  [5,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [6,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#>  [7,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#>  [8,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [9,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [10,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [11,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [12,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [13,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [14,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [15,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [16,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [17,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [18,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [19,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [20,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [21,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [22,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [23,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [24,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [25,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [26,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [27,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [28,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [29,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [30,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [31,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [32,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [33,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [34,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [35,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [36,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [37,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [38,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [39,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [40,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#>  [2,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#>  [3,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#>  [4,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#>  [5,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [6,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#>  [7,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#>  [8,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [9,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [10,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [11,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [12,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [13,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [14,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [15,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [16,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [17,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [18,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [19,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [20,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [21,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [22,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [23,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [24,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [25,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [26,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [27,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [28,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [29,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [30,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [31,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [32,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [33,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [34,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [35,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [36,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [37,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [38,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [39,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [40,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#>  [2,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#>  [3,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#>  [4,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#>  [5,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [6,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#>  [7,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#>  [8,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [9,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [10,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [11,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [12,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [13,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [14,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [15,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [16,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [17,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [18,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [19,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [20,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [21,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [22,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [23,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [24,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [25,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [26,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [27,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [28,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [29,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [30,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [31,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [32,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [33,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [34,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [35,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [36,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [37,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [38,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [39,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [40,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8763003 0.7870848 0.5041056 0.6542658 0.8898759 0.6974148 0.1901143
#>  [2,] 0.7196703 0.5867776 0.3936011 0.5248663 0.7779408 0.5402081 0.1637752
#>  [3,] 0.8917952 0.7923986 0.5682593 0.6605594 0.9058794 0.7578430 0.1517035
#>  [4,] 0.9080248 0.8275008 0.6544000 0.7210371 0.9229020 0.8023822 0.2086669
#>  [5,] 0.8477486 0.7758473 0.6264146 0.7290852 0.8596513 0.7738606 0.3390448
#>  [6,] 0.8288535 0.7329448 0.4388669 0.5318229 0.8911619 0.6664869 0.1287016
#>  [7,] 0.8475047 0.7580903 0.6235727 0.7312566 0.8600224 0.7644470 0.3227239
#>  [8,] 0.9071658 0.8278018 0.6568651 0.7557737 0.9224097 0.8176620 0.2494511
#>  [9,] 0.8724604 0.8035149 0.5227254 0.5749222 0.9187338 0.6398321 0.1551368
#> [10,] 0.9009969 0.8349837 0.5814684 0.6878880 0.9230627 0.7873916 0.1864874
#> [11,] 0.9264260 0.8644313 0.6255819 0.7174620 0.9414152 0.7880632 0.1941869
#> [12,] 0.8716466 0.7777824 0.6157872 0.7065208 0.8912451 0.7725277 0.2465567
#> [13,] 0.8360116 0.7895118 0.5244713 0.5667193 0.8884580 0.6040118 0.1977436
#> [14,] 0.8360236 0.7347163 0.4658883 0.5731591 0.8731760 0.6747825 0.1608078
#> [15,] 0.8865087 0.8117724 0.5737471 0.7110347 0.8966093 0.7788356 0.2173153
#> [16,] 0.8804622 0.8129967 0.5565995 0.6694447 0.8997644 0.6871717 0.2083706
#> [17,] 0.8679546 0.7876809 0.5719788 0.7210299 0.8842255 0.7580973 0.2896630
#> [18,] 0.9129900 0.7993658 0.5780861 0.7450845 0.9113755 0.7754381 0.2075062
#> [19,] 0.8729032 0.8057748 0.5872009 0.6643213 0.8958469 0.7279851 0.2248971
#> [20,] 0.8763805 0.7812153 0.5618572 0.6184047 0.9029205 0.7448111 0.1600516
#> [21,] 0.8645088 0.7645580 0.6094458 0.6996981 0.8875146 0.7688493 0.2408661
#> [22,] 0.8786583 0.8254619 0.5659719 0.6619126 0.9035001 0.7081905 0.2377590
#> [23,] 0.7784146 0.7202850 0.4765164 0.5123795 0.8464127 0.5230244 0.1927597
#> [24,] 0.8876759 0.8297829 0.5962315 0.6872455 0.9034506 0.7750060 0.2272018
#> [25,] 0.8681600 0.7762272 0.6206403 0.7158440 0.8912048 0.7794473 0.2559907
#> [26,] 0.9229718 0.8769706 0.6564584 0.7365403 0.9372862 0.8069548 0.2454820
#> [27,] 0.7350146 0.5611171 0.3564890 0.5001804 0.8133266 0.5161693 0.1290213
#> [28,] 0.9017572 0.8141693 0.5699552 0.6656155 0.9306250 0.7144222 0.1644689
#> [29,] 0.8033316 0.7399338 0.5084417 0.6108709 0.8596410 0.6498759 0.2311560
#> [30,] 0.7621835 0.6085121 0.4573606 0.6334762 0.7827871 0.6428697 0.2818908
#> [31,] 0.8891469 0.8168109 0.5677020 0.6642560 0.9125514 0.7589320 0.1702104
#> [32,] 0.8520653 0.7861579 0.5411392 0.6716048 0.8736015 0.7133259 0.2491747
#> [33,] 0.8252904 0.7829656 0.4575367 0.6140999 0.8590957 0.6472453 0.2073663
#> [34,] 0.8880672 0.8153606 0.6251540 0.6961498 0.9110118 0.7771451 0.2230310
#> [35,] 0.8825562 0.7871139 0.5809567 0.6960859 0.9085976 0.7806100 0.2117047
#> [36,] 0.8092736 0.6946428 0.4474785 0.6072297 0.8457244 0.6584761 0.1862382
#> [37,] 0.7664967 0.6839244 0.5099306 0.5699313 0.8179628 0.6327941 0.2461389
#> [38,] 0.8522293 0.7824309 0.6245216 0.7307029 0.8644501 0.7724536 0.3254500
#> [39,] 0.8783318 0.7629685 0.6250710 0.7461593 0.8821837 0.8131386 0.2802930
#> [40,] 0.8902639 0.8441361 0.6166280 0.6977745 0.8993175 0.7388912 0.2482437
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6000853 0.5880474 0.4520013
#>  [2,] 0.3302560 0.5486769 0.2788176
#>  [3,] 0.5582276 0.6385466 0.4695883
#>  [4,] 0.6993163 0.6780720 0.6083500
#>  [5,] 0.6654583 0.6350377 0.5841256
#>  [6,] 0.5156156 0.5117486 0.3419477
#>  [7,] 0.6633845 0.6297494 0.5853967
#>  [8,] 0.6898501 0.6922661 0.6295478
#>  [9,] 0.6559753 0.6248508 0.4737306
#> [10,] 0.6115723 0.6529972 0.4584327
#> [11,] 0.7175388 0.6817407 0.5993360
#> [12,] 0.6490675 0.6413586 0.5614041
#> [13,] 0.6079344 0.6249491 0.4496862
#> [14,] 0.4239327 0.6075288 0.3414721
#> [15,] 0.6117519 0.6474708 0.5146375
#> [16,] 0.5951286 0.6592898 0.5225481
#> [17,] 0.6074146 0.6299797 0.5189063
#> [18,] 0.5879592 0.6670691 0.5361705
#> [19,] 0.6438362 0.6345092 0.5493188
#> [20,] 0.5902839 0.6423139 0.4702136
#> [21,] 0.6376089 0.6350857 0.5322686
#> [22,] 0.6357222 0.6542888 0.5351724
#> [23,] 0.5453751 0.5653728 0.4017489
#> [24,] 0.6350646 0.6397213 0.5140566
#> [25,] 0.6428311 0.6447235 0.5529917
#> [26,] 0.7384371 0.7041105 0.6252205
#> [27,] 0.3083596 0.5322039 0.2398401
#> [28,] 0.6810718 0.6495747 0.5209334
#> [29,] 0.5863264 0.6193462 0.5130921
#> [30,] 0.3968117 0.5807819 0.3419713
#> [31,] 0.6366366 0.6398108 0.5202490
#> [32,] 0.5841129 0.6089286 0.4937673
#> [33,] 0.5311455 0.5657494 0.3958174
#> [34,] 0.6999961 0.6578647 0.5305443
#> [35,] 0.6220948 0.6636305 0.4966913
#> [36,] 0.4616495 0.5703388 0.3923662
#> [37,] 0.5418951 0.5547869 0.4094294
#> [38,] 0.6736892 0.6375579 0.5899920
#> [39,] 0.6016355 0.6614188 0.5069702
#> [40,] 0.6672905 0.6863680 0.5958209
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4727453 0.5272547
#>  [2,] 0.4750088 0.5249912
#>  [3,] 0.4758746 0.5241254
#>  [4,] 0.4746863 0.5253137
#>  [5,] 0.4669024 0.5330976
#>  [6,] 0.4748171 0.5251829
#>  [7,] 0.4675420 0.5324580
#>  [8,] 0.4724245 0.5275755
#>  [9,] 0.4780190 0.5219810
#> [10,] 0.4726772 0.5273228
#> [11,] 0.4753964 0.5246036
#> [12,] 0.4715598 0.5284402
#> [13,] 0.4761821 0.5238179
#> [14,] 0.4751222 0.5248778
#> [15,] 0.4722300 0.5277700
#> [16,] 0.4751338 0.5248662
#> [17,] 0.4679837 0.5320163
#> [18,] 0.4727494 0.5272506
#> [19,] 0.4739488 0.5260512
#> [20,] 0.4766248 0.5233752
#> [21,] 0.4711120 0.5288880
#> [22,] 0.4736590 0.5263410
#> [23,] 0.4763022 0.5236978
#> [24,] 0.4720560 0.5279440
#> [25,] 0.4705722 0.5294278
#> [26,] 0.4732687 0.5267313
#> [27,] 0.4756030 0.5243970
#> [28,] 0.4760017 0.5239983
#> [29,] 0.4747314 0.5252686
#> [30,] 0.4672353 0.5327647
#> [31,] 0.4758071 0.5241929
#> [32,] 0.4709721 0.5290279
#> [33,] 0.4720411 0.5279589
#> [34,] 0.4721791 0.5278209
#> [35,] 0.4723219 0.5276781
#> [36,] 0.4733412 0.5266588
#> [37,] 0.4712247 0.5287753
#> [38,] 0.4677280 0.5322720
#> [39,] 0.4676361 0.5323639
#> [40,] 0.4743520 0.5256480
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]         [,4]     [,5]       [,6]
#>  [1,] 1.9578519 1.3074418  0.01642279 0.6378436086 2.089474 0.83501723
#>  [2,] 0.9428265 0.3506598 -0.43219992 0.0995472881 1.253707 0.16118046
#>  [3,] 2.1092104 1.3394444  0.27475245 0.6657882159 2.264330 1.14089011
#>  [4,] 2.2897531 1.5680172  0.63843688 0.9496116927 2.482446 1.40125050
#>  [5,] 1.7170508 1.2416285  0.51686572 0.9899863029 1.812397 1.23024005
#>  [6,] 1.5775238 1.0096151 -0.24576177 0.1274639126 2.102665 0.69233821
#>  [7,] 1.7151627 1.1422378  0.50474062 1.0010076327 1.815476 1.17721654
#>  [8,] 2.2795107 1.5701276  0.64935521 1.1296464952 2.475547 1.50058674
#>  [9,] 1.9228900 1.4084091  0.09096421 0.3019625534 2.425266 0.57463539
#> [10,] 2.2083509 1.6213683  0.32880426 0.7902639488 2.484707 1.30927353
#> [11,] 2.5330422 1.8525934  0.51330934 0.9319067210 2.776909 1.31329030
#> [12,] 1.9155969 1.2527899  0.47170470 0.8785458085 2.103523 1.22263938
#> [13,] 1.6288466 1.3219851  0.09796346 0.2684782783 2.075086 0.42220946
#> [14,] 1.6289346 1.0186844 -0.13665900 0.2947521226 1.929337 0.72989618
#> [15,] 2.0555644 1.4615684  0.29715595 0.9004146919 2.160105 1.25889376
#> [16,] 1.9968140 1.4696006  0.22737251 0.7056744485 2.194609 0.78692987
#> [17,] 1.8829936 1.3110026  0.28992916 0.9495759903 2.033068 1.14227635
#> [18,] 2.3507020 1.3823352  0.31492136 1.0725654630 2.330547 1.23927703
#> [19,] 1.9268757 1.4227859  0.35240615 0.6826116425 2.151907 0.98442381
#> [20,] 1.9585919 1.2727625  0.24870273 0.4827826753 2.230104 1.07112648
#> [21,] 1.8532546 1.1778332  0.44498308 0.8458607336 2.065602 1.20182537
#> [22,] 1.9797853 1.5538000  0.26543506 0.6718289761 2.236735 0.88661211
#> [23,] 1.2564514 0.9458757 -0.09400348 0.0495280476 1.706738 0.09216262
#> [24,] 2.0672188 1.5840893  0.38978722 0.7872728912 2.236166 1.23679693
#> [25,] 1.8847871 1.2438138  0.49226694 0.9239387380 2.103107 1.26244825
#> [26,] 2.4834272 1.9640502  0.64755095 1.0280633243 2.704406 1.43034305
#> [27,] 1.0202158 0.2456970 -0.59063556 0.0007215162 1.471772 0.06469967
#> [28,] 2.2169037 1.4773324  0.28166826 0.6884205558 2.596330 0.91695968
#> [29,] 1.4072487 1.0456243  0.03377020 0.4509745446 1.812311 0.61849369
#> [30,] 1.1646879 0.4410624 -0.17097307 0.5471588075 1.281983 0.58784114
#> [31,] 2.0820563 1.4948886  0.27248156 0.6823186202 2.345193 1.14683307
#> [32,] 1.7508924 1.3019198  0.16492957 0.7154525233 1.933185 0.91159229
#> [33,] 1.5526102 1.2830331 -0.17026324 0.4645791723 1.807800 0.60695238
#> [34,] 2.0711488 1.4852257  0.51148281 0.8290299537 2.326052 1.24910622
#> [35,] 2.0168627 1.3076155  0.32670192 0.8287278219 2.296631 1.26922462
#> [36,] 1.4452970 0.8219156 -0.21086368 0.4356823147 1.701453 0.65651056
#> [37,] 1.1886341 0.7718662  0.03972779 0.2815706972 1.502606 0.54422224
#> [38,] 1.7521938 1.2798889  0.50878477 0.9981917906 1.852754 1.22221754
#> [39,] 1.9767266 1.1690239  0.51112861 1.0782323542 2.013274 1.47053442
#> [40,] 2.0934397 1.6893307  0.47526028 0.8367227863 2.189664 1.04021354
#>             [,7]       [,8]       [,9]       [,10]
#>  [1,] -1.4492676  0.4058207 0.35589926 -0.19258772
#>  [2,] -1.6304026 -0.7070276 0.19532616 -0.95033444
#>  [3,] -1.7213019  0.2339721 0.56906163 -0.12179695
#>  [4,] -1.3329796  0.8440445 0.74492549  0.44038181
#>  [5,] -0.6675537  0.6877144 0.55389042  0.33973273
#>  [6,] -1.9124883  0.0624826 0.04700323 -0.65462673
#>  [7,] -0.7412819  0.6784136 0.53114197  0.34496756
#>  [8,] -1.1015419  0.7994185 0.81073505  0.53027754
#>  [9,] -1.6948671  0.6454098 0.51018889 -0.10517450
#> [10,] -1.4729978  0.4539261 0.63223983 -0.16665386
#> [11,] -1.4230310  0.9322857 0.76178272  0.40269910
#> [12,] -1.1170620  0.6149429 0.58126577  0.24686263
#> [13,] -1.4004572  0.4386380 0.51060831 -0.20193849
#> [14,] -1.6522301 -0.3066498 0.43693663 -0.65674077
#> [15,] -1.2813804  0.4546821 0.60794031  0.05856661
#> [16,] -1.3347751  0.3852079 0.66013103  0.09025378
#> [17,] -0.8970213  0.4364576 0.53212961  0.07566121
#> [18,] -1.3400235  0.3555353 0.69495876  0.14493536
#> [19,] -1.2373530  0.5920537 0.55161078  0.19791867
#> [20,] -1.6578445  0.3651393 0.58542135 -0.11928685
#> [21,] -1.1479371  0.5650009 0.55409721  0.12925399
#> [22,] -1.1650050  0.5568448 0.63794518  0.14092236
#> [23,] -1.4321769  0.1820010 0.26299686 -0.39818350
#> [24,] -1.2241796  0.5540064 0.57415480  0.05624132
#> [25,] -1.0669125  0.5876731 0.59592521  0.21276583
#> [26,] -1.1228556  1.0378611 0.86694946  0.51176659
#> [27,] -1.9096402 -0.8077995 0.12899415 -1.15355646
#> [28,] -1.6253462  0.7587020 0.61717047  0.08378257
#> [29,] -1.2017951  0.3487995 0.48677427  0.05238050
#> [30,] -0.9351016 -0.4187677 0.32598401 -0.65452173
#> [31,] -1.5841365  0.5607956 0.57454295  0.08104015
#> [32,] -1.1030187  0.3396804 0.44281101 -0.02493189
#> [33,] -1.3408742  0.1247433 0.26452938 -0.42292347
#> [34,] -1.2480897  0.8472793 0.65379295  0.12232933
#> [35,] -1.3146803  0.4984493 0.67951528 -0.01323519
#> [36,] -1.4746410 -0.1537039 0.28323339 -0.43737681
#> [37,] -1.1193119  0.1679744 0.22003098 -0.36632458
#> [38,] -0.7288371  0.7249185 0.56478020  0.36393241
#> [39,] -0.9430088  0.4122844 0.66962326  0.02788259
#> [40,] -1.1080015  0.6959555 0.78319362  0.38808207
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] 0.09472569 0.2038528
#>  [2,] 0.07183596 0.1718843
#>  [3,] 0.10447287 0.2010496
#>  [4,] 0.11265823 0.2139998
#>  [5,] 0.08814404 0.2207285
#>  [6,] 0.08715968 0.1879765
#>  [7,] 0.08832962 0.2183443
#>  [8,] 0.10770832 0.2181222
#>  [9,] 0.11335754 0.2013381
#> [10,] 0.10172016 0.2111204
#> [11,] 0.11651492 0.2150090
#> [12,] 0.09812044 0.2120043
#> [13,] 0.10836269 0.2037066
#> [14,] 0.09024885 0.1898424
#> [15,] 0.09797151 0.2091660
#> [16,] 0.10831770 0.2078647
#> [17,] 0.08803933 0.2162799
#> [18,] 0.10057350 0.2096839
#> [19,] 0.10529971 0.2095991
#> [20,] 0.10510341 0.1986725
#> [21,] 0.09517689 0.2108579
#> [22,] 0.10624568 0.2117075
#> [23,] 0.09952553 0.1943879
#> [24,] 0.10162464 0.2135174
#> [25,] 0.09579538 0.2136426
#> [26,] 0.11515771 0.2221851
#> [27,] 0.07073515 0.1684006
#> [28,] 0.11150904 0.2075760
#> [29,] 0.09528910 0.1964498
#> [30,] 0.06070722 0.1919539
#> [31,] 0.10677346 0.2036207
#> [32,] 0.09236972 0.2086119
#> [33,] 0.08795987 0.1999122
#> [34,] 0.10400787 0.2154065
#> [35,] 0.09725620 0.2080819
#> [36,] 0.08117651 0.1879128
#> [37,] 0.08129098 0.1965195
#> [38,] 0.09052178 0.2197893
#> [39,] 0.08726464 0.2169017
#> [40,] 0.11264181 0.2153240
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4727453  0.4727453
#>  [2,]  0.5249912 -0.5249912
#>  [3,]  0.5241254 -0.5241254
#>  [4,]  0.5253137 -0.5253137
#>  [5,] -0.4669024  0.4669024
#>  [6,]  0.5251829 -0.5251829
#>  [7,] -0.4675420  0.4675420
#>  [8,]  0.5275755 -0.5275755
#>  [9,]  0.5219810 -0.5219810
#> [10,]  0.5273228 -0.5273228
#> [11,]  0.5246036 -0.5246036
#> [12,]  0.5284402 -0.5284402
#> [13,]  0.5238179 -0.5238179
#> [14,] -0.4751222  0.4751222
#> [15,] -0.4722300  0.4722300
#> [16,] -0.4751338  0.4751338
#> [17,] -0.4679837  0.4679837
#> [18,]  0.5272506 -0.5272506
#> [19,] -0.4739488  0.4739488
#> [20,]  0.5233752 -0.5233752
#> [21,]  0.5288880 -0.5288880
#> [22,]  0.5263410 -0.5263410
#> [23,]  0.5236978 -0.5236978
#> [24,]  0.5279440 -0.5279440
#> [25,]  0.5294278 -0.5294278
#> [26,] -0.4732687  0.4732687
#> [27,]  0.5243970 -0.5243970
#> [28,] -0.4760017  0.4760017
#> [29,]  0.5252686 -0.5252686
#> [30,] -0.4672353  0.4672353
#> [31,]  0.5241929 -0.5241929
#> [32,] -0.4709721  0.4709721
#> [33,] -0.4720411  0.4720411
#> [34,] -0.4721791  0.4721791
#> [35,]  0.5276781 -0.5276781
#> [36,]  0.5266588 -0.5266588
#> [37,]  0.5287753 -0.5287753
#> [38,] -0.4677280  0.4677280
#> [39,] -0.4676361  0.4676361
#> [40,] -0.4743520  0.4743520
#> 
#> $L
#> [1] 0.6906024 0.6881702 0.6928052 0.6997019 0.7464336 0.7001486
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5072464