Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13          V14
#>  [1,] -0.2976406 -0.20421141 -0.143872137 -0.167769607 -0.02622712 -0.165133132
#>  [2,] -0.3265659 -0.27168998 -0.092098382 -0.170960494 -0.18626528 -0.194242851
#>  [3,] -0.4626861 -0.02653792 -0.019382468  0.042706309 -0.05437393 -0.118475157
#>  [4,] -0.2913523 -0.10089270 -0.171985054  0.003469767 -0.04817822 -0.057416285
#>  [5,] -0.3974292 -0.06082571 -0.049914780 -0.157059730 -0.10153648 -0.046841337
#>  [6,] -0.2080920 -0.03773736 -0.155737000 -0.130116674 -0.02027509  0.003428059
#>  [7,] -0.2088276 -0.09570360 -0.066643550 -0.149415614 -0.10805362 -0.098675684
#>  [8,] -0.2436952 -0.10537187 -0.157102235 -0.074223378 -0.10812641 -0.025907484
#>  [9,] -0.3603877 -0.25647232 -0.150483254 -0.136062670 -0.27786296 -0.147084427
#> [10,] -0.2444774 -0.10998913 -0.009093198 -0.171573022 -0.21975735 -0.113871991
#>               V15          V16         V17         V18         V19         V2
#>  [1,] -0.09078670 -0.009986127 -0.07966130  0.03092043 -0.09625801 -0.1939372
#>  [2,] -0.08502523 -0.008531406  0.01325193 -0.07709902  0.02072059 -0.3778875
#>  [3,] -0.01424428  0.022575941  0.17772805  0.18083358  0.09513288 -0.3211230
#>  [4,] -0.18329270 -0.159758271 -0.11140981  0.01495357 -0.04127389 -0.3324478
#>  [5,] -0.14607534 -0.054925343 -0.03591517  0.07609994  0.05008712 -0.3327439
#>  [6,] -0.02330388  0.028356809  0.07126011 -0.03329155  0.01498381 -0.2603052
#>  [7,]  0.06372334  0.085335542  0.07620203  0.01884841  0.14213729 -0.2086541
#>  [8,] -0.08069020  0.032780124 -0.05383676  0.06756332  0.01779374 -0.4378004
#>  [9,] -0.06217937 -0.130921571 -0.05512052 -0.08043563 -0.17623366 -0.3036306
#> [10,] -0.01600819 -0.131423641  0.05858455 -0.10042063  0.09494349 -0.3285110
#>               V20         V21          V22         V23         V24        V25
#>  [1,]  0.06984017  0.07201862 -0.009777904  0.21365567  0.17766593 0.21301276
#>  [2,] -0.09805661 -0.03740045 -0.024202214  0.07827098  0.01185304 0.09148596
#>  [3,]  0.11348934  0.30553509  0.256326411  0.25477350  0.21749926 0.18947623
#>  [4,] -0.10770134  0.01991315 -0.081104853 -0.01399420  0.13264513 0.03144681
#>  [5,]  0.02130374  0.04521662  0.036105275  0.08082393  0.14559754 0.10221845
#>  [6,]  0.05470683  0.08739541  0.098228800  0.06458852  0.01480417 0.15776710
#>  [7,]  0.07050017  0.03535636  0.044465094  0.02987506  0.15537310 0.07990927
#>  [8,]  0.18938673  0.21459216  0.225690993  0.14774731  0.09227506 0.23053340
#>  [9,] -0.02203613 -0.02054715 -0.024930594 -0.04045901 -0.02016601 0.00322384
#> [10,]  0.03572012  0.01045992  0.117136163  0.21021267  0.15767439 0.04272379
#>                V26          V27          V28          V29         V3
#>  [1,]  0.161548806  0.079487614  0.208498175  0.096728953 -0.3020369
#>  [2,]  0.110230899 -0.004417865  0.006478946  0.004778537 -0.1994554
#>  [3,]  0.312997080  0.315459461  0.179613875  0.280263967 -0.4283796
#>  [4,]  0.009428281  0.078739861  0.062504005  0.023571704 -0.2955102
#>  [5,]  0.192889717  0.152172554  0.139053505  0.240276304 -0.2547439
#>  [6,] -0.037718709  0.021479330  0.113186114 -0.003384093 -0.3301067
#>  [7,]  0.210939685  0.152612054  0.194892244  0.085559954 -0.2907688
#>  [8,]  0.278166261  0.359988797  0.303134291  0.246982472 -0.2656051
#>  [9,]  0.034324506 -0.016983218 -0.032735257 -0.139471423 -0.3498043
#> [10,]  0.168318556  0.192807625  0.202033726  0.177830529 -0.2688946
#>                V30         V31         V32         V33           V34
#>  [1,]  0.094669534  0.11029579  0.09751798 -0.07427237  0.0006927715
#>  [2,]  0.001813592  0.01927704 -0.02268054 -0.07666336 -0.0300180624
#>  [3,]  0.226943578  0.09164334  0.06082580  0.15443154  0.1441683152
#>  [4,]  0.032887946 -0.05996565 -0.07355219 -0.05186191 -0.0047026509
#>  [5,]  0.030124405  0.10697962  0.02658138  0.02552550 -0.0799048017
#>  [6,] -0.028608342 -0.06370049 -0.05317221 -0.05065012 -0.0412149649
#>  [7,]  0.149662465  0.08545714 -0.05231639 -0.08547859 -0.0772610693
#>  [8,]  0.153916246  0.13106919  0.13685121  0.12962708  0.1188112980
#>  [9,]  0.003829250 -0.14928006 -0.07947781 -0.15392666 -0.2584743361
#> [10,]  0.003130973  0.08316371 -0.09847920  0.03179870 -0.1264380294
#>                V35          V36          V37         V38         V39         V4
#>  [1,] -0.023634104 -0.042945419  0.074125907  0.01086849  0.07975344 -0.3382013
#>  [2,]  0.004981582 -0.031479504 -0.056678136 -0.03690594 -0.14917837 -0.2996207
#>  [3,] -0.017678330  0.042724982  0.130390757  0.09902903 -0.02908575 -0.3160906
#>  [4,] -0.055513351 -0.084575374 -0.144076181 -0.12926990 -0.04797533 -0.3586680
#>  [5,] -0.081079681 -0.004187511 -0.098960542 -0.23025681 -0.10547275 -0.2166820
#>  [6,] -0.033493976 -0.121598656  0.060640006  0.03751067 -0.09693283 -0.2025088
#>  [7,] -0.068250775  0.023062522  0.007309264 -0.07598839  0.05377707 -0.1357383
#>  [8,]  0.179416966  0.064566507 -0.040567011 -0.02836468 -0.04246037 -0.2644874
#>  [9,] -0.224562275 -0.110566060 -0.171437723 -0.18240767 -0.23708587 -0.2951274
#> [10,] -0.121234018  0.081603214 -0.046146132 -0.00732190 -0.01367377 -0.2746058
#>                V40          V41          V42         V43         V44
#>  [1,] -0.105088838 -0.045895305 -0.086790723 -0.13058129 -0.05981593
#>  [2,] -0.128882088 -0.017229893 -0.022979046 -0.15078218 -0.12233164
#>  [3,]  0.001146223 -0.046082816 -0.085171320  0.04099188 -0.10744251
#>  [4,] -0.063747994 -0.001704541 -0.045938840 -0.10512187 -0.14172068
#>  [5,] -0.056457403 -0.074368888 -0.151730533 -0.01606035 -0.17397625
#>  [6,] -0.095582737  0.010930768 -0.028092110 -0.04616502 -0.06684489
#>  [7,] -0.020579195  0.058846691  0.037858819 -0.12577274 -0.20841960
#>  [8,]  0.046577734 -0.125694379  0.001554167 -0.15335619 -0.08672801
#>  [9,] -0.289663717 -0.208846109 -0.132962772 -0.14554896 -0.28002517
#> [10,] -0.089341307 -0.157043275 -0.156605336 -0.02638573 -0.17585148
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.18942306 -0.14104923 -0.1138739 -0.2002387 -0.1965653 -0.2056020
#>  [2,] -0.24340465 -0.07803403 -0.2320277 -0.2154387 -0.3283394 -0.2284689
#>  [3,] -0.18860635 -0.13454492 -0.2015841 -0.3077926 -0.3155617 -0.2295695
#>  [4,] -0.19423262 -0.25935997 -0.1366326 -0.3037204 -0.2416084 -0.2537485
#>  [5,] -0.12412157 -0.22276360 -0.1244263 -0.2982666 -0.2354356 -0.2973592
#>  [6,] -0.13546785 -0.12446756 -0.1988567 -0.1985104 -0.2789791 -0.1499225
#>  [7,] -0.04675497 -0.06263146 -0.2500463 -0.2297081 -0.3032631 -0.2688890
#>  [8,] -0.04292150 -0.17039101 -0.1651672 -0.2245619 -0.3485753 -0.2665850
#>  [9,] -0.17649384 -0.12579603 -0.2612809 -0.2017551 -0.2844271 -0.3170365
#> [10,] -0.09827322 -0.24038319 -0.2521891 -0.2948949 -0.2166923 -0.2160628
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2606083 -0.2835671 -0.3892364 -0.2927201 -0.2752206 -0.2613023
#>  [2,] -0.3122570 -0.3606923 -0.2172688 -0.2554773 -0.2430265 -0.3603290
#>  [3,] -0.3900636 -0.3779602 -0.5192336 -0.4085643 -0.3959825 -0.4758228
#>  [4,] -0.2964834 -0.2687385 -0.2693539 -0.2545500 -0.4225081 -0.2679438
#>  [5,] -0.3637311 -0.2858414 -0.3918082 -0.3744688 -0.3866440 -0.3585964
#>  [6,] -0.3808667 -0.3772971 -0.2644105 -0.3533140 -0.3526056 -0.2456470
#>  [7,] -0.2479165 -0.3167436 -0.1999246 -0.2611316 -0.2784814 -0.3707806
#>  [8,] -0.4176818 -0.3865293 -0.3063090 -0.4846422 -0.3090366 -0.3235032
#>  [9,] -0.2367809 -0.2495652 -0.3223664 -0.2501192 -0.3492333 -0.2447613
#> [10,] -0.3698113 -0.4059241 -0.3864697 -0.3961051 -0.3194858 -0.3027040
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3880499 -0.4085855 -0.3842775 -0.3203414 -0.2450116 -0.3067370
#>  [2,] -0.3127928 -0.2748513 -0.2237770 -0.2926506 -0.3294088 -0.2699689
#>  [3,] -0.4049034 -0.5174246 -0.4510060 -0.3881180 -0.1941744 -0.4765043
#>  [4,] -0.3180971 -0.3917618 -0.3501805 -0.3978358 -0.1470723 -0.2651660
#>  [5,] -0.3829784 -0.3200393 -0.2955389 -0.3136178 -0.2091248 -0.2672588
#>  [6,] -0.3395777 -0.3148925 -0.2883974 -0.3843854 -0.1508445 -0.3046601
#>  [7,] -0.3596083 -0.3888830 -0.4000511 -0.2396511 -0.1259537 -0.2992920
#>  [8,] -0.3489602 -0.3651600 -0.3849257 -0.3927012 -0.3289901 -0.4076094
#>  [9,] -0.2912881 -0.2425139 -0.3237877 -0.2750566 -0.1755083 -0.3262595
#> [10,] -0.3892070 -0.2979041 -0.3966305 -0.3124395 -0.1327983 -0.3389236
#>               V7          V8          V9
#>  [1,] -0.1736217 -0.18607023 -0.10732222
#>  [2,] -0.1373139 -0.14121374 -0.06342453
#>  [3,] -0.2104264 -0.24850071 -0.18606799
#>  [4,] -0.2701836 -0.17120890 -0.15000784
#>  [5,] -0.2572828 -0.23438648 -0.16688373
#>  [6,] -0.2768607 -0.13078469 -0.13011454
#>  [7,] -0.2149031 -0.22395556 -0.09103048
#>  [8,] -0.1540458 -0.08567963 -0.19408413
#>  [9,] -0.2645153 -0.17794040 -0.18581850
#> [10,] -0.1580292 -0.27506996 -0.07718279
#> 
#> $W[[2]]
#>             [,1]         [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] -0.02383291 -0.005419435  0.1080414 0.03957219 -0.00119113 0.05156722
#> [2,]  0.07407931  0.048024760 -0.1102008 0.09964107  0.01701047 0.02141271
#>             [,7]        [,8]       [,9]       [,10]
#> [1,]  0.09643278  0.04657270 0.05003250 -0.02176130
#> [2,] -0.08524291 -0.06796155 0.09154026  0.04496873
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  4.913737e-04  0.0029106964  0.0035158725  0.0035997856  3.242523e-03
#>  [2,]  1.727317e-04  0.0010512981  0.0012581106  0.0014138959  1.341756e-03
#>  [3,] -4.018475e-05 -0.0001241293 -0.0002293787 -0.0001173457 -2.461014e-05
#>  [4,]  2.367940e-04  0.0014648468  0.0018034123  0.0019718988  1.819169e-03
#>  [5,]  2.977329e-04  0.0018118978  0.0021554906  0.0022947703  2.148540e-03
#>  [6,]  1.698538e-04  0.0010672232  0.0012153996  0.0013480316  1.336848e-03
#>  [7,] -9.437444e-05 -0.0004710541 -0.0007789358 -0.0005897257 -3.009701e-04
#>  [8,]  6.246300e-05  0.0004286127  0.0004078349  0.0004871768  5.178049e-04
#>  [9,]  3.357684e-05  0.0002061155  0.0002706839  0.0003074580  2.712229e-04
#> [10,]  4.248679e-04  0.0025533951  0.0031147239  0.0032344337  2.928724e-03
#>                 V14          V15          V16          V17          V18
#>  [1,]  3.093869e-03 2.837623e-03 0.0033108143 0.0034880455 0.0040434792
#>  [2,]  1.273725e-03 1.157622e-03 0.0014192621 0.0015840873 0.0019217902
#>  [3,]  3.839020e-05 5.570831e-05 0.0001331136 0.0002143942 0.0003188686
#>  [4,]  1.669212e-03 1.494086e-03 0.0017930325 0.0019998192 0.0024141120
#>  [5,]  2.086783e-03 1.931498e-03 0.0023146268 0.0025092008 0.0029583765
#>  [6,]  1.360225e-03 1.297357e-03 0.0015983167 0.0017668335 0.0021084948
#>  [7,] -5.089269e-05 5.731403e-05 0.0002475508 0.0004141525 0.0005769929
#>  [8,]  6.046895e-04 6.087534e-04 0.0007872864 0.0008662014 0.0010150731
#>  [9,]  2.182117e-04 1.902005e-04 0.0002256756 0.0002607763 0.0003195684
#> [10,]  2.750984e-03 2.507553e-03 0.0029336285 0.0031334583 0.0036785231
#>                V19            V2           V20           V21          V22
#>  [1,] 5.359650e-03  5.072931e-04  6.847972e-03  7.383508e-03 0.0065882522
#>  [2,] 2.311893e-03  1.746783e-04  2.802225e-03  3.129854e-03 0.0030858189
#>  [3,] 1.407239e-04 -4.140631e-05 -1.919056e-05  5.405795e-05 0.0003242638
#>  [4,] 3.032422e-03  2.394051e-04  3.711391e-03  4.084412e-03 0.0039338803
#>  [5,] 3.692826e-03  3.098948e-04  4.537681e-03  4.945737e-03 0.0046484642
#>  [6,] 2.442676e-03  1.775242e-04  2.903011e-03  3.220745e-03 0.0032007945
#>  [7,] 3.968926e-05 -9.141255e-05 -3.670665e-04 -2.236969e-04 0.0003787829
#>  [8,] 1.038050e-03  6.511009e-05  1.182752e-03  1.350799e-03 0.0014103460
#>  [9,] 4.042024e-04  3.188254e-05  4.929170e-04  5.474183e-04 0.0005455200
#> [10,] 4.843021e-03  4.395926e-04  6.100347e-03  6.585890e-03 0.0059867059
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0060624626 0.0059677457 0.0061275245 0.0063216303 0.0059065345
#>  [2,] 0.0031366436 0.0031997287 0.0033639618 0.0033882479 0.0031664391
#>  [3,] 0.0005831353 0.0006327053 0.0006472010 0.0006392763 0.0006442349
#>  [4,] 0.0038774492 0.0038971074 0.0040597821 0.0041096311 0.0038649756
#>  [5,] 0.0045141148 0.0045016463 0.0046385821 0.0047197868 0.0044303487
#>  [6,] 0.0032953919 0.0033629250 0.0035070828 0.0035477031 0.0033497707
#>  [7,] 0.0009782264 0.0011244637 0.0011887629 0.0010972632 0.0010891296
#>  [8,] 0.0015233207 0.0015460752 0.0015866021 0.0015961894 0.0015203470
#>  [9,] 0.0005430141 0.0005410793 0.0005647341 0.0005672773 0.0005381716
#> [10,] 0.0055968799 0.0055185693 0.0056679623 0.0058091549 0.0054385597
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0055385291 0.0046776960  5.224565e-04 0.0039366071 0.0025763388
#>  [2,] 0.0029493910 0.0025497102  2.009390e-04 0.0021398319 0.0017167842
#>  [3,] 0.0006380497 0.0006307408 -2.465614e-05 0.0005667015 0.0007403388
#>  [4,] 0.0035685806 0.0030331948  2.690160e-04 0.0024897300 0.0018456307
#>  [5,] 0.0041688785 0.0035795823  3.380773e-04 0.0030303477 0.0022663890
#>  [6,] 0.0031928427 0.0028277926  2.076388e-04 0.0024617483 0.0020726589
#>  [7,] 0.0011638458 0.0012583826 -5.114996e-05 0.0012663003 0.0017800612
#>  [8,] 0.0014749042 0.0013232488  8.353545e-05 0.0012023450 0.0011590563
#>  [9,] 0.0004827294 0.0003988497  3.766937e-05 0.0003164913 0.0002341957
#> [10,] 0.0050741300 0.0042750000  4.625227e-04 0.0035487151 0.0023653458
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0023522319 0.0020399763 0.0018815107 1.156429e-03 1.052350e-03
#>  [2,] 0.0014554611 0.0012505141 0.0011382565 1.006797e-03 1.042851e-03
#>  [3,] 0.0005868743 0.0005586133 0.0005261922 7.002992e-04 7.966838e-04
#>  [4,] 0.0015831176 0.0013418949 0.0011759746 8.645225e-04 8.850520e-04
#>  [5,] 0.0020006429 0.0017551795 0.0016321901 1.309119e-03 1.321651e-03
#>  [6,] 0.0017657228 0.0015854379 0.0015024979 1.448857e-03 1.491168e-03
#>  [7,] 0.0014270446 0.0013546267 0.0013849192 1.943232e-03 2.096541e-03
#>  [8,] 0.0009752266 0.0008796504 0.0008430742 9.164258e-04 9.985175e-04
#>  [9,] 0.0001936976 0.0001479313 0.0001027905 3.443318e-05 3.177345e-05
#> [10,] 0.0021435305 0.0018409863 0.0016505817 9.974958e-04 9.340049e-04
#>                V37          V38          V39            V4          V40
#>  [1,] 1.753616e-03 0.0024555616 0.0025983130  6.982207e-04 0.0016196607
#>  [2,] 1.160163e-03 0.0012522991 0.0012690784  2.387443e-04 0.0009929026
#>  [3,] 6.334389e-04 0.0004429913 0.0004117267 -5.379289e-05 0.0005476065
#>  [4,] 1.174932e-03 0.0014126257 0.0014906635  3.359526e-04 0.0010872435
#>  [5,] 1.639561e-03 0.0019246269 0.0019629451  4.316915e-04 0.0014230358
#>  [6,] 1.510760e-03 0.0015216564 0.0014812573  2.521644e-04 0.0012342835
#>  [7,] 1.536115e-03 0.0009780563 0.0007482979 -1.187064e-04 0.0010736746
#>  [8,] 9.659194e-04 0.0009026110 0.0008443556  9.429948e-05 0.0007661949
#>  [9,] 9.575861e-05 0.0001466366 0.0001699300  4.595589e-05 0.0001197982
#> [10,] 1.570912e-03 0.0021649103 0.0023119212  6.109530e-04 0.0014952995
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0013842522 0.0022109609  2.722577e-03  2.397802e-03  0.0022999655
#>  [2,] 0.0008730232 0.0010321130  1.185485e-03  9.714131e-04  0.0007915584
#>  [3,] 0.0004321226 0.0001879034  7.155784e-05 -2.421669e-06 -0.0001477052
#>  [4,] 0.0009592604 0.0012928732  1.532203e-03  1.299537e-03  0.0011232300
#>  [5,] 0.0012293078 0.0015898205  1.850915e-03  1.575702e-03  0.0014038591
#>  [6,] 0.0010692890 0.0011226213  1.213554e-03  9.894548e-04  0.0007975788
#>  [7,] 0.0008820577 0.0002308643 -3.519007e-05 -1.939051e-04 -0.0004996869
#>  [8,] 0.0006274931 0.0005302029  5.161400e-04  3.924334e-04  0.0002686621
#>  [9,] 0.0001097108 0.0001736039  2.124439e-04  1.786199e-04  0.0001558105
#> [10,] 0.0013126343 0.0020193021  2.445880e-03  2.135003e-03  0.0020212150
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0016253940  9.735028e-04  8.524137e-04  5.086437e-04  9.306141e-04
#>  [2,]  0.0005692079  3.870748e-04  3.071167e-04  1.850527e-04  3.597607e-04
#>  [3,] -0.0001204961 -3.877341e-05 -5.828712e-05 -3.102404e-05 -1.746419e-05
#>  [4,]  0.0007967389  5.400345e-04  4.402125e-04  2.636356e-04  4.919968e-04
#>  [5,]  0.0010001005  6.344093e-04  5.279092e-04  3.176759e-04  6.055061e-04
#>  [6,]  0.0005773840  3.890756e-04  3.039979e-04  1.814269e-04  3.755123e-04
#>  [7,] -0.0003470203 -1.537134e-04 -1.929014e-04 -1.106993e-04 -8.306720e-05
#>  [8,]  0.0001872575  1.301597e-04  9.814655e-05  6.003480e-05  1.547734e-04
#>  [9,]  0.0001139489  8.947217e-05  7.016482e-05  4.047735e-05  6.917117e-05
#> [10,]  0.0014399187  8.963941e-04  7.679769e-04  4.589156e-04  8.325381e-04
#>                V50           V51           V52          V53           V54
#>  [1,] 1.534516e-04  1.529774e-04  1.445536e-04 1.158934e-04  1.504760e-04
#>  [2,] 7.421840e-05  5.727253e-05  5.442312e-05 5.043912e-05  6.445599e-05
#>  [3,] 1.086872e-05 -9.488730e-06 -7.434529e-06 1.416669e-06 -3.989170e-07
#>  [4,] 8.887818e-05  7.831591e-05  7.164216e-05 6.302593e-05  8.105553e-05
#>  [5,] 1.116300e-04  9.772175e-05  9.337674e-05 7.899087e-05  1.005677e-04
#>  [6,] 8.249225e-05  5.919098e-05  5.888509e-05 5.387394e-05  6.538641e-05
#>  [7,] 2.562123e-05 -2.210366e-05 -1.309170e-05 4.157040e-06 -2.456422e-06
#>  [8,] 3.946603e-05  2.085096e-05  2.361686e-05 2.299029e-05  2.894637e-05
#>  [9,] 1.197075e-05  1.105366e-05  9.217115e-06 7.849664e-06  1.078752e-05
#> [10,] 1.376821e-04  1.361603e-04  1.272012e-04 1.021888e-04  1.324227e-04
#>                V55           V56          V57           V58           V59
#>  [1,] 7.468939e-05  8.607003e-05 3.775660e-05  8.221542e-05  5.948221e-05
#>  [2,] 3.367567e-05  3.309059e-05 2.111396e-05  3.038913e-05  2.072883e-05
#>  [3,] 2.744455e-06 -2.412156e-06 7.336093e-06 -4.659051e-06 -2.824558e-06
#>  [4,] 4.081803e-05  4.319032e-05 2.239842e-05  4.067038e-05  2.782877e-05
#>  [5,] 5.312696e-05  5.557844e-05 3.165838e-05  5.187796e-05  3.809691e-05
#>  [6,] 3.801544e-05  3.530609e-05 2.796596e-05  3.188038e-05  2.450897e-05
#>  [7,] 7.821980e-06 -6.101557e-06 2.327723e-05 -9.412256e-06 -4.107067e-06
#>  [8,] 1.781424e-05  1.385387e-05 1.670812e-05  1.189827e-05  9.872176e-06
#>  [9,] 5.051085e-06  5.557552e-06 1.916607e-06  5.643013e-06  3.333364e-06
#> [10,] 6.637347e-05  7.516626e-05 3.296651e-05  7.175305e-05  5.133800e-05
#>                 V6          V60           V7            V8            V9
#>  [1,] 1.131897e-03 4.233309e-05 1.096623e-03  1.521154e-03  0.0025709271
#>  [2,] 5.221693e-04 1.893826e-05 5.226668e-04  5.623315e-04  0.0008852000
#>  [3,] 4.173017e-05 2.019070e-06 6.613228e-05 -2.927363e-05 -0.0001457712
#>  [4,] 6.725558e-04 2.287707e-05 6.572892e-04  7.506460e-04  0.0012330896
#>  [5,] 7.908297e-04 3.034362e-05 7.887275e-04  9.714787e-04  0.0015704493
#>  [6,] 5.365052e-04 2.265037e-05 5.585099e-04  6.090508e-04  0.0009140404
#>  [7,] 3.667125e-05 6.907861e-06 1.091374e-04 -1.166953e-04 -0.0004509337
#>  [8,] 2.409527e-04 9.961995e-06 2.588195e-04  2.849387e-04  0.0003760453
#>  [9,] 1.004876e-04 2.721727e-06 9.608667e-05  1.018530e-04  0.0001639476
#> [10,] 1.030541e-03 3.752122e-05 1.001024e-03  1.323882e-03  0.0022265643
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.05116722 -0.01391259 -0.08206461 -0.02195113 -0.04771901 -0.04221745
#> [2,]  0.05116722  0.01391259  0.08206461  0.02195113  0.04771901  0.04221745
#>             [,7]        [,8]         [,9]     [,10]
#> [1,] -0.06126138 -0.07523591 -0.003042315 -0.049536
#> [2,]  0.06126138  0.07523591  0.003042315  0.049536
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.4095815  0.1570475  0.8809253  0.3943204  0.4038281  0.2576603
#>  [7]  0.6388166  0.6806173 -0.3715523  0.3092752
#> 
#> $B[[2]]
#> [1]  0.06363169 -0.08524239
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0076249844 0.0040904395 0.0010557199 0.0048209333 0.0058670005
#>  [6] 0.0045767097 0.0021595181 0.0022998957 0.0006124988 0.0069463191
#> 
#> $vB[[2]]
#> [1] -0.08791846  0.08791846
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#>  [2,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#>  [3,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#>  [4,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#>  [5,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#>  [6,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#>  [7,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#>  [8,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#>  [9,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [10,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [11,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [12,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [13,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [14,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [15,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [16,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [17,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [18,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [19,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [20,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [21,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [22,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [23,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [24,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [25,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [26,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [27,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [28,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [29,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [30,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [31,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [32,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [33,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [34,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [35,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [36,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [37,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [38,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [39,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [40,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#>  [2,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#>  [3,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#>  [4,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#>  [5,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#>  [6,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#>  [7,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#>  [8,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#>  [9,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [10,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [11,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [12,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [13,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [14,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [15,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [16,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [17,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [18,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [19,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [20,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [21,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [22,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [23,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [24,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [25,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [26,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [27,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [28,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [29,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [30,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [31,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [32,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [33,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [34,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [35,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [36,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [37,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [38,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [39,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [40,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#>  [2,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#>  [3,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#>  [4,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#>  [5,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#>  [6,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#>  [7,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#>  [8,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#>  [9,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [10,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [11,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [12,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [13,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [14,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [15,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [16,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [17,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [18,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [19,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [20,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [21,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [22,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [23,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [24,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [25,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [26,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [27,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [28,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [29,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [30,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [31,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [32,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [33,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [34,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [35,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [36,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [37,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [38,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [39,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [40,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#>  [2,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#>  [3,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#>  [4,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#>  [5,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#>  [6,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#>  [7,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#>  [8,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#>  [9,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [10,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [11,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [12,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [13,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [14,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [15,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [16,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [17,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [18,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [19,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [20,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [21,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [22,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [23,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [24,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [25,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [26,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [27,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [28,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [29,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [30,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [31,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [32,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [33,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [34,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [35,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [36,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [37,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [38,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [39,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [40,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#>  [2,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#>  [3,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#>  [4,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#>  [5,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#>  [6,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#>  [7,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#>  [8,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#>  [9,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [10,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [11,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [12,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [13,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [14,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [15,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [16,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [17,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [18,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [19,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [20,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [21,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [22,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [23,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [24,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [25,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [26,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [27,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [28,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [29,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [30,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [31,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [32,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [33,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [34,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [35,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [36,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [37,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [38,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [39,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [40,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#>  [2,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#>  [3,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#>  [4,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#>  [5,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#>  [6,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#>  [7,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#>  [8,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#>  [9,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [10,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [11,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [12,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [13,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [14,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [15,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [16,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [17,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [18,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [19,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [20,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [21,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [22,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [23,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [24,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [25,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [26,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [27,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [28,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [29,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [30,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [31,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [32,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [33,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [34,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [35,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [36,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [37,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [38,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [39,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [40,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6029977 0.2786645 0.8834887 0.3019377 0.5066221 0.4272533 0.6873323
#>  [2,] 0.7149531 0.4188727 0.9403539 0.4450753 0.6478865 0.5502817 0.7431921
#>  [3,] 0.7602269 0.3968918 0.9617001 0.3898145 0.6453585 0.5069278 0.7786662
#>  [4,] 0.7636475 0.4037221 0.9628459 0.4523099 0.7331170 0.5800783 0.8175917
#>  [5,] 0.7488478 0.4310893 0.9346616 0.4975532 0.7018352 0.5383276 0.7781605
#>  [6,] 0.6759132 0.3933754 0.9365753 0.4028999 0.6138062 0.5601685 0.7543155
#>  [7,] 0.7501811 0.4083920 0.9501089 0.4771733 0.7228099 0.5705331 0.7904907
#>  [8,] 0.7665506 0.4335985 0.9612199 0.4274496 0.6931333 0.5495278 0.7967799
#>  [9,] 0.6742178 0.3599947 0.9231183 0.3861468 0.5671580 0.5011190 0.7047709
#> [10,] 0.7394266 0.3739123 0.9592516 0.4239667 0.6865383 0.5765575 0.8053457
#> [11,] 0.7027850 0.3540884 0.9554297 0.3807247 0.6496819 0.5549529 0.7938968
#> [12,] 0.6513875 0.3296325 0.8924101 0.3649488 0.5695802 0.4523223 0.6834249
#> [13,] 0.7478553 0.3835736 0.9645536 0.4197818 0.7137920 0.5902287 0.8167477
#> [14,] 0.6826566 0.3546684 0.9429328 0.4195220 0.6464443 0.5106837 0.7556562
#> [15,] 0.6185659 0.3331991 0.9172217 0.3556870 0.5370931 0.4563951 0.7069770
#> [16,] 0.7036569 0.3787143 0.9469789 0.3936943 0.6192941 0.5742661 0.7682758
#> [17,] 0.7725379 0.4623360 0.9427249 0.5126000 0.7112009 0.5353219 0.7769122
#> [18,] 0.7551997 0.4006159 0.9599298 0.4368306 0.6699602 0.5618871 0.7946635
#> [19,] 0.7893314 0.4513362 0.9520952 0.5154069 0.7529436 0.5544223 0.7923112
#> [20,] 0.7068025 0.3399552 0.9470339 0.4005029 0.6456628 0.5369577 0.7484531
#> [21,] 0.6704469 0.3614924 0.9306489 0.3675089 0.6110889 0.5268715 0.7747477
#> [22,] 0.6309373 0.3163387 0.8979454 0.3626429 0.5715305 0.4652222 0.7312906
#> [23,] 0.6960490 0.3328682 0.9470148 0.3913147 0.6340175 0.5465166 0.7437395
#> [24,] 0.6898209 0.3830298 0.9283940 0.4200573 0.6474004 0.5384898 0.7652746
#> [25,] 0.7694973 0.4493922 0.9467282 0.4984080 0.7306861 0.5582918 0.7900046
#> [26,] 0.6559743 0.4016983 0.8831956 0.4123421 0.5614818 0.4624268 0.6728020
#> [27,] 0.5103572 0.2360638 0.8051344 0.2646103 0.3522966 0.3559303 0.5795043
#> [28,] 0.6567974 0.3657621 0.8961766 0.4155697 0.5493247 0.4566874 0.6994352
#> [29,] 0.4912881 0.2138646 0.8720784 0.2323121 0.3833410 0.3996853 0.6239454
#> [30,] 0.8158378 0.4679313 0.9709591 0.4936963 0.7443531 0.5840276 0.8137464
#> [31,] 0.6401525 0.2628561 0.9326675 0.3514602 0.5890239 0.4725154 0.7124166
#> [32,] 0.5141463 0.2137127 0.8633798 0.2132439 0.3638135 0.3811444 0.6205294
#> [33,] 0.7244920 0.3780919 0.9466028 0.4518444 0.6875079 0.5258413 0.7710104
#> [34,] 0.7558791 0.4725667 0.9284275 0.5141518 0.6902460 0.5800456 0.7759983
#> [35,] 0.7859706 0.4771752 0.9677491 0.4883309 0.7622038 0.6440898 0.8440700
#> [36,] 0.5176294 0.2439905 0.8425487 0.2639305 0.3870953 0.4199427 0.6066895
#> [37,] 0.5756153 0.2764727 0.9170604 0.3166179 0.5306169 0.5061837 0.7048112
#> [38,] 0.7582592 0.4354119 0.9419007 0.4800996 0.7150210 0.5303553 0.7748611
#> [39,] 0.5929265 0.3391938 0.9076675 0.3645281 0.5333845 0.5448342 0.6933776
#> [40,] 0.7729622 0.4011846 0.9670580 0.4291863 0.6832364 0.5649804 0.7966457
#>            [,8]       [,9]     [,10]
#>  [1,] 0.8297863 0.05916526 0.4802421
#>  [2,] 0.8961181 0.11634081 0.6368883
#>  [3,] 0.9329252 0.06406501 0.6589306
#>  [4,] 0.9333712 0.11994472 0.7315150
#>  [5,] 0.9059007 0.14132136 0.6780216
#>  [6,] 0.8719933 0.11885541 0.6130311
#>  [7,] 0.9172487 0.14336960 0.7140732
#>  [8,] 0.9350225 0.09710422 0.6896959
#>  [9,] 0.8878801 0.07677547 0.6021247
#> [10,] 0.9189665 0.10370410 0.6955747
#> [11,] 0.9092108 0.08511923 0.6516206
#> [12,] 0.8196208 0.08976376 0.5727648
#> [13,] 0.9250532 0.11032628 0.7145859
#> [14,] 0.9116581 0.08149427 0.6274553
#> [15,] 0.8783491 0.06077738 0.5353373
#> [16,] 0.8908318 0.09336656 0.6411162
#> [17,] 0.9238620 0.13014410 0.6968412
#> [18,] 0.9312548 0.08588455 0.7033175
#> [19,] 0.9264550 0.15137719 0.7356748
#> [20,] 0.9040596 0.09422484 0.6592250
#> [21,] 0.8704920 0.08533491 0.6185601
#> [22,] 0.8605908 0.08085543 0.5497095
#> [23,] 0.8964068 0.09131112 0.6529374
#> [24,] 0.8806581 0.11575080 0.6338245
#> [25,] 0.9172203 0.16914676 0.7245836
#> [26,] 0.8537524 0.09235996 0.5274786
#> [27,] 0.7180815 0.03506141 0.3453614
#> [28,] 0.8691207 0.06787782 0.5401148
#> [29,] 0.8023736 0.03411700 0.3969795
#> [30,] 0.9542446 0.11027511 0.7593973
#> [31,] 0.8803929 0.06182666 0.6083136
#> [32,] 0.8104466 0.03431857 0.4112878
#> [33,] 0.9140011 0.09945449 0.6612257
#> [34,] 0.8934017 0.17300151 0.6985831
#> [35,] 0.9395127 0.16917086 0.7675223
#> [36,] 0.7790361 0.05569392 0.4160051
#> [37,] 0.8268455 0.06863824 0.5327687
#> [38,] 0.9179717 0.12177047 0.6915081
#> [39,] 0.8146148 0.09785733 0.5235809
#> [40,] 0.9399826 0.08250381 0.7171546
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4240157 0.5759843
#>  [2,] 0.4007280 0.5992720
#>  [3,] 0.4011773 0.5988227
#>  [4,] 0.3930311 0.6069689
#>  [5,] 0.3946636 0.6053364
#>  [6,] 0.4062076 0.5937924
#>  [7,] 0.3933855 0.6066145
#>  [8,] 0.3959978 0.6040022
#>  [9,] 0.4093184 0.5906816
#> [10,] 0.3982998 0.6017002
#> [11,] 0.4046541 0.5953459
#> [12,] 0.4130458 0.5869542
#> [13,] 0.3962402 0.6037598
#> [14,] 0.4055876 0.5944124
#> [15,] 0.4176417 0.5823583
#> [16,] 0.4043933 0.5956067
#> [17,] 0.3916526 0.6083474
#> [18,] 0.3969924 0.6030076
#> [19,] 0.3881639 0.6118361
#> [20,] 0.4034669 0.5965331
#> [21,] 0.4088882 0.5911118
#> [22,] 0.4155690 0.5844310
#> [23,] 0.4048312 0.5951688
#> [24,] 0.4038521 0.5961479
#> [25,] 0.3903308 0.6096692
#> [26,] 0.4115193 0.5884807
#> [27,] 0.4397688 0.5602312
#> [28,] 0.4127086 0.5872914
#> [29,] 0.4392235 0.5607765
#> [30,] 0.3868071 0.6131929
#> [31,] 0.4137534 0.5862466
#> [32,] 0.4385582 0.5614418
#> [33,] 0.3996800 0.6003200
#> [34,] 0.3917086 0.6082914
#> [35,] 0.3864509 0.6135491
#> [36,] 0.4346726 0.5653274
#> [37,] 0.4217997 0.5782003
#> [38,] 0.3939400 0.6060600
#> [39,] 0.4177639 0.5822361
#> [40,] 0.3954544 0.6045456
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]     [,3]         [,4]        [,5]         [,6]
#>  [1,]  0.41797131 -0.95109587 2.025890 -0.838087534  0.02648976 -0.293066700
#>  [2,]  0.91956344 -0.32740291 2.757827 -0.220588986  0.60976200  0.201808982
#>  [3,]  1.15392404 -0.41843292 3.223255 -0.448091869  0.59869871  0.027712820
#>  [4,]  1.17278213 -0.38998016 3.254818 -0.191342017  1.01049489  0.323094860
#>  [5,]  1.09247647 -0.27740822 2.660605 -0.009787121  0.85605229  0.153611929
#>  [6,]  0.73505311 -0.43314560 2.692377 -0.393396810  0.46334001  0.241845873
#>  [7,]  1.09957843 -0.37061664 2.946734 -0.091370174  0.95844281  0.284026336
#>  [8,]  1.18893532 -0.26718410 3.210295 -0.292264512  0.81480875  0.198763083
#>  [9,]  0.72732415 -0.57538698 2.485490 -0.463538022  0.27026522  0.004475908
#> [10,]  1.04299021 -0.51546926 3.158737 -0.306510499  0.78398469  0.308657158
#> [11,]  0.86059545 -0.60111601 3.065093 -0.486473511  0.61764134  0.220703157
#> [12,]  0.62514361 -0.70984788 2.115599 -0.553948426  0.28013870 -0.191291975
#> [13,]  1.08720661 -0.47440756 3.303644 -0.323669324  0.91387298  0.364911074
#> [14,]  0.76600761 -0.59858109 2.804765 -0.324736137  0.60344606  0.042741440
#> [15,]  0.48346580 -0.69375149 2.405184 -0.594133604  0.14864532 -0.174863898
#> [16,]  0.86477320 -0.49500914 2.882586 -0.431809428  0.48655328  0.299278436
#> [17,]  1.22269716 -0.15094205 2.800908  0.050410685  0.90122373  0.141523451
#> [18,]  1.12653968 -0.40289964 3.176226 -0.254034927  0.70800486  0.248824322
#> [19,]  1.32089982 -0.19527318 2.989449  0.061647194  1.11437373  0.218555176
#> [20,]  0.87990471 -0.66349409 2.883683 -0.403370183  0.60002832  0.148100983
#> [21,]  0.71020682 -0.56889250 2.596700 -0.542918780  0.45189153  0.107589621
#> [22,]  0.53624009 -0.77064935 2.174601 -0.563911369  0.28809824 -0.139336019
#> [23,]  0.82855343 -0.69524117 2.883303 -0.441789136  0.54949099  0.186606104
#> [24,]  0.79928201 -0.47670790 2.562277 -0.322538139  0.60763192  0.154264570
#> [25,]  1.20547507 -0.20312694 2.877606 -0.006367950  0.99810646  0.234232323
#> [26,]  0.64540540 -0.39839378 2.023046 -0.354291607  0.24717808 -0.150576810
#> [27,]  0.04143491 -1.17438225 1.418699 -1.022142321 -0.60895937 -0.593072165
#> [28,]  0.64905454 -0.55044101 2.155446 -0.340987239  0.19794261 -0.173685870
#> [29,] -0.03485101 -1.30178612 1.919462 -1.195301360 -0.47539145 -0.406776711
#> [30,]  1.48839866 -0.12845100 3.509579 -0.025216228  1.06871849  0.339329140
#> [31,]  0.57602627 -1.03117615 2.628406 -0.612627073  0.35993170 -0.110049503
#> [32,]  0.05660048 -1.30268971 1.843650 -1.305481983 -0.55885054 -0.484693845
#> [33,]  0.96685421 -0.49765485 2.875120 -0.193221118  0.78849425  0.103457223
#> [34,]  1.13021796 -0.10984362 2.562782  0.056622296  0.80126980  0.322960451
#> [35,]  1.30080617 -0.09136252 3.401426 -0.046684682  1.16479969  0.593159502
#> [36,]  0.07054674 -1.13092462 1.677315 -1.025638619 -0.45953851 -0.323008504
#> [37,]  0.30479933 -0.96202598 2.403061 -0.769358662  0.12262107  0.024736201
#> [38,]  1.14315899 -0.25980394 2.785745 -0.079643716  0.91989645  0.121570524
#> [39,]  0.37607676 -0.66688915 2.285482 -0.555764270  0.13373702  0.179819667
#> [40,]  1.22511346 -0.40053152 3.379511 -0.285171788  0.76868535  0.261399878
#>            [,7]      [,8]      [,9]       [,10]
#>  [1,] 0.7876768 1.5841136 -2.766433 -0.07907258
#>  [2,] 1.0626263 2.1548172 -2.027548  0.56188396
#>  [3,] 1.2579106 2.6325168 -2.681648  0.65853225
#>  [4,] 1.5001158 2.6396653 -1.992954  1.00232278
#>  [5,] 1.2549786 2.2645794 -1.804358  0.74469470
#>  [6,] 1.1217628 1.9186989 -2.003314  0.46007127
#>  [7,] 1.3278856 2.4055391 -1.787581  0.91524958
#>  [8,] 1.3662888 2.6665298 -2.229822  0.79869814
#>  [9,] 0.8701212 2.0692682 -2.486987  0.41432605
#> [10,] 1.4200464 2.4283864 -2.156729  0.82631257
#> [11,] 1.3485762 2.3040365 -2.374741  0.62617055
#> [12,] 0.7695561 1.5137802 -2.316523  0.29314052
#> [13,] 1.4944664 2.5130721 -2.087413  0.91776199
#> [14,] 1.1290104 2.3340504 -2.422215  0.52131520
#> [15,] 0.8807469 1.9768883 -2.737835  0.14158511
#> [16,] 1.1986007 2.0992659 -2.273205  0.58021189
#> [17,] 1.2477619 2.4960159 -1.899685  0.83230061
#> [18,] 1.3532687 2.6061261 -2.364953  0.86314609
#> [19,] 1.3389134 2.5334682 -1.723840  1.02360814
#> [20,] 1.0903790 2.2431684 -2.263107  0.65984263
#> [21,] 1.2353164 1.9053160 -2.371974  0.48344097
#> [22,] 1.0011804 1.8202054 -2.430781  0.19949718
#> [23,] 1.0654965 2.1579227 -2.297730  0.63197615
#> [24,] 1.1818185 1.9986771 -2.033299  0.54865930
#> [25,] 1.3249532 2.4051649 -1.591686  0.96731306
#> [26,] 0.7208855 1.7643405 -2.285154  0.11002523
#> [27,] 0.3207389 0.9349651 -3.314964 -0.63949209
#> [28,] 0.8446099 1.8932064 -2.619755  0.16080462
#> [29,] 0.5063285 1.4011961 -3.343247 -0.41806653
#> [30,] 1.4745397 3.0376104 -2.087934  1.14937782
#> [31,] 0.9071499 1.9961557 -2.719600  0.44022884
#> [32,] 0.4917959 1.4529149 -3.337147 -0.35864423
#> [33,] 1.2140251 2.3634975 -2.203301  0.66876112
#> [34,] 1.2424965 2.1259691 -1.564503  0.84055961
#> [35,] 1.6888283 2.7429272 -1.591515  1.19437329
#> [36,] 0.4334177 1.2600578 -2.830579 -0.33919491
#> [37,] 0.8703147 1.5634334 -2.607798  0.13126311
#> [38,] 1.2359660 2.4151023 -1.975770  0.80717950
#> [39,] 0.8159576 1.4802798 -2.221262  0.09439343
#> [40,] 1.3654601 2.7512273 -2.408804  0.93039076
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.09848204 0.2078281
#>  [2,] -0.11976716 0.2826655
#>  [3,] -0.12727405 0.2732880
#>  [4,] -0.13133297 0.3032559
#>  [5,] -0.12532835 0.3024224
#>  [6,] -0.11252642 0.2671388
#>  [7,] -0.12803791 0.3050655
#>  [8,] -0.12980962 0.2923596
#>  [9,] -0.11406693 0.2527167
#> [10,] -0.12550619 0.2870481
#> [11,] -0.11949799 0.2666122
#> [12,] -0.10960426 0.2417842
#> [13,] -0.12828401 0.2928719
#> [14,] -0.11775863 0.2644780
#> [15,] -0.10465181 0.2278101
#> [16,] -0.11760143 0.2695914
#> [17,] -0.13014649 0.3102245
#> [18,] -0.12840477 0.2896079
#> [19,] -0.13406726 0.3209693
#> [20,] -0.12093230 0.2701081
#> [21,] -0.11366375 0.2548995
#> [22,] -0.10595078 0.2350392
#> [23,] -0.11912728 0.2662476
#> [24,] -0.11661679 0.2728231
#> [25,] -0.13004694 0.3158750
#> [26,] -0.10825630 0.2494318
#> [27,] -0.07862901 0.1634714
#> [28,] -0.10824340 0.2445359
#> [29,] -0.08128672 0.1630272
#> [30,] -0.13923701 0.3215166
#> [31,] -0.11199438 0.2364758
#> [32,] -0.08414060 0.1628750
#> [33,] -0.12364186 0.2831569
#> [34,] -0.12526523 0.3148706
#> [35,] -0.13398809 0.3282676
#> [36,] -0.08432943 0.1784825
#> [37,] -0.09829165 0.2170981
#> [38,] -0.12929903 0.3014815
#> [39,] -0.09799976 0.2339599
#> [40,] -0.13164376 0.2927978
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5759843 -0.5759843
#>  [2,]  0.5992720 -0.5992720
#>  [3,] -0.4011773  0.4011773
#>  [4,]  0.6069689 -0.6069689
#>  [5,]  0.6053364 -0.6053364
#>  [6,] -0.4062076  0.4062076
#>  [7,]  0.6066145 -0.6066145
#>  [8,] -0.3959978  0.3959978
#>  [9,] -0.4093184  0.4093184
#> [10,]  0.6017002 -0.6017002
#> [11,]  0.5953459 -0.5953459
#> [12,]  0.5869542 -0.5869542
#> [13,]  0.6037598 -0.6037598
#> [14,]  0.5944124 -0.5944124
#> [15,] -0.4176417  0.4176417
#> [16,]  0.5956067 -0.5956067
#> [17,] -0.3916526  0.3916526
#> [18,]  0.6030076 -0.6030076
#> [19,]  0.6118361 -0.6118361
#> [20,]  0.5965331 -0.5965331
#> [21,] -0.4088882  0.4088882
#> [22,] -0.4155690  0.4155690
#> [23,]  0.5951688 -0.5951688
#> [24,] -0.4038521  0.4038521
#> [25,]  0.6096692 -0.6096692
#> [26,] -0.4115193  0.4115193
#> [27,] -0.4397688  0.4397688
#> [28,] -0.4127086  0.4127086
#> [29,]  0.5607765 -0.5607765
#> [30,] -0.3868071  0.3868071
#> [31,]  0.5862466 -0.5862466
#> [32,]  0.5614418 -0.5614418
#> [33,]  0.6003200 -0.6003200
#> [34,] -0.3917086  0.3917086
#> [35,] -0.3864509  0.3864509
#> [36,]  0.5653274 -0.5653274
#> [37,]  0.5782003 -0.5782003
#> [38,] -0.3939400  0.3939400
#> [39,]  0.5822361 -0.5822361
#> [40,]  0.6045456 -0.6045456
#> 
#> $L
#> [1] 0.6854696 0.7056378 0.7222274 0.6734930 0.8197306 0.7469547
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609