Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11          V12          V13
#>  [1,] -0.2892237 -0.071424352 -0.144068064 -0.118188718 -0.093405681
#>  [2,] -0.2828999 -0.005138728 -0.072944076 -0.119834403 -0.105755363
#>  [3,] -0.2261067 -0.207125147 -0.164702835 -0.059741168 -0.007714852
#>  [4,] -0.3452613 -0.063570644 -0.078881222 -0.034406037 -0.067491489
#>  [5,] -0.2821493 -0.081414439  0.003192577 -0.092191450 -0.065176039
#>  [6,] -0.1994809 -0.036345170 -0.142504505 -0.151115772 -0.008489276
#>  [7,] -0.2850069 -0.129688377 -0.053180332 -0.087510162 -0.074115954
#>  [8,] -0.3976707 -0.157552870 -0.066604404 -0.005705265  0.021320870
#>  [9,] -0.3460200 -0.192065264 -0.062012095 -0.165011300 -0.137176482
#> [10,] -0.2085105 -0.084583991 -0.049058991 -0.047224802 -0.183827953
#>               V14          V15         V16          V17         V18         V19
#>  [1,] -0.10852501 -0.152304052 -0.18031598 -0.196047399 -0.15589115 -0.03194095
#>  [2,] -0.14138836 -0.075166769 -0.02686890  0.014356347  0.01275539  0.06935916
#>  [3,] -0.01935980  0.064414736  0.07373549 -0.041832250  0.06420228  0.04520750
#>  [4,]  0.02508416 -0.046544962 -0.01992331  0.010561789  0.10098161  0.04816654
#>  [5,] -0.06328754 -0.102356760  0.10920917 -0.016308822  0.04653702  0.09616248
#>  [6,] -0.01950714  0.050374310  0.01868095  0.005838284  0.16012478  0.26825316
#>  [7,] -0.03012960 -0.029710080 -0.03793501  0.105413881  0.01108975  0.06889224
#>  [8,] -0.03210130  0.008024099  0.10469649  0.162210752  0.14661935  0.15533694
#>  [9,] -0.02594190 -0.006049371 -0.02054316 -0.046277166 -0.05586195 -0.04955411
#> [10,] -0.01344702 -0.035868856 -0.09145488 -0.028449948 -0.02669039  0.04920723
#>               V2          V20          V21          V22          V23        V24
#>  [1,] -0.2071988 -0.103138315  0.033204831 -0.056329360 -0.064564413 0.12191619
#>  [2,] -0.2200037  0.076314894  0.021866964  0.094628306  0.004106528 0.01648638
#>  [3,] -0.2995542 -0.018976847  0.087523649  0.034533942  0.063046626 0.06547015
#>  [4,] -0.1850018  0.112919867  0.096169221  0.116957904  0.224232045 0.11678834
#>  [5,] -0.3563987  0.102709010  0.145306369  0.137284315  0.101098169 0.12883221
#>  [6,] -0.2452448  0.196538543  0.274191390  0.217683171  0.033086339 0.08871194
#>  [7,] -0.2849529  0.160436675  0.091811969  0.097955428  0.157105057 0.19336510
#>  [8,] -0.3187980  0.180447323  0.198965281  0.222040670  0.224708971 0.28239476
#>  [9,] -0.3913868 -0.004595965 -0.020188241  0.005779548  0.095402808 0.07079717
#> [10,] -0.2628234  0.022085278  0.003328353  0.039064701 -0.019988535 0.09808649
#>              V25        V26        V27         V28         V29         V3
#>  [1,] 0.01232943 0.10511962 0.05276499  0.05582484  0.02884803 -0.2287598
#>  [2,] 0.05185625 0.04457559 0.04026313  0.11544491  0.11222380 -0.1456661
#>  [3,] 0.11101124 0.24637876 0.15843338  0.16466876  0.19893046 -0.2105935
#>  [4,] 0.13402961 0.17951293 0.26817238  0.24180846  0.08138703 -0.2720441
#>  [5,] 0.01580976 0.04102229 0.22170652 -0.03812104  0.07070940 -0.3554158
#>  [6,] 0.13575784 0.02076727 0.00974407  0.03650040 -0.02575943 -0.3342277
#>  [7,] 0.21595511 0.25014189 0.15877295  0.16700916  0.20892625 -0.3063114
#>  [8,] 0.15902182 0.15281606 0.14310546  0.18204824  0.20898778 -0.3357117
#>  [9,] 0.19838042 0.24137132 0.17684508  0.22404156  0.23759925 -0.2418505
#> [10,] 0.01453790 0.18978859 0.10117825  0.08614627  0.01817458 -0.1957109
#>                V30         V31         V32         V33          V34
#>  [1,] -0.115694257 -0.10819129 -0.07267176 -0.14809008 -0.058431561
#>  [2,]  0.009572787  0.01952998 -0.05687628  0.10805908 -0.035780143
#>  [3,]  0.048389721  0.11650508  0.02905484  0.02882762 -0.055001754
#>  [4,]  0.096380177  0.03491428  0.01491412 -0.02687526  0.010420917
#>  [5,] -0.021987834 -0.02599945 -0.07872962 -0.12631194 -0.016505280
#>  [6,]  0.095217033  0.09165847  0.05334373  0.02569327  0.090059974
#>  [7,]  0.185220600  0.09994322  0.13427326  0.05877327  0.009726042
#>  [8,]  0.139232846 -0.06269017 -0.05307378 -0.08288128  0.004544509
#>  [9,]  0.150231105  0.06130253  0.07485150  0.01757463  0.019996907
#> [10,] -0.015667469 -0.13505591 -0.18292281 -0.08631267 -0.228315066
#>                V35          V36         V37          V38         V39         V4
#>  [1,] -0.067381679 -0.194966575 -0.05294996 -0.091237560 -0.10623689 -0.2432605
#>  [2,]  0.017755058 -0.112014728 -0.05430233 -0.039760623  0.06121179 -0.2204066
#>  [3,] -0.010536478 -0.005562711 -0.06603357 -0.137063200 -0.18609678 -0.2566854
#>  [4,]  0.020433970  0.039460310 -0.05089584 -0.070256694  0.01489004 -0.2239219
#>  [5,] -0.136114421 -0.212525686 -0.14735163 -0.020011227 -0.03628248 -0.3037099
#>  [6,]  0.039483353 -0.107889047  0.05010601  0.060885687 -0.07193351 -0.1845480
#>  [7,] -0.077717788  0.006938439 -0.12299470 -0.122833016 -0.13629219 -0.2942093
#>  [8,]  0.076164929  0.031013989 -0.04325137  0.001389184 -0.05556523 -0.4188664
#>  [9,] -0.001819063 -0.060756922  0.05416106 -0.062431718 -0.06356129 -0.2553701
#> [10,] -0.142600544 -0.129155152 -0.06588702 -0.201100092 -0.20069353 -0.2436649
#>               V40         V41          V42          V43         V44         V45
#>  [1,] -0.08022062 -0.23361259 -0.098619594 -0.199268140 -0.08753207 -0.10730922
#>  [2,]  0.03533127 -0.07566093 -0.111659242 -0.180159486 -0.07689359 -0.08807890
#>  [3,] -0.07632742 -0.07380397 -0.178424947 -0.110148398 -0.22838209 -0.11906072
#>  [4,] -0.08086762 -0.03177521 -0.054266149 -0.125600459 -0.01812254 -0.15432117
#>  [5,] -0.12054740 -0.08859004 -0.121163216 -0.035923247 -0.24235859 -0.22257451
#>  [6,] -0.07281495  0.05186109 -0.020302281  0.002601854 -0.16026613 -0.09080281
#>  [7,] -0.07920427  0.02779651 -0.031621360 -0.143498054 -0.08606269 -0.01483041
#>  [8,] -0.08033607 -0.01911581 -0.005307951 -0.014215205 -0.05794566 -0.10369280
#>  [9,] -0.07405705 -0.10003478 -0.025316431 -0.151533059 -0.17619142 -0.08614878
#> [10,] -0.06288804 -0.21994577 -0.245001006 -0.195327320 -0.11006689 -0.24184057
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.29318529 -0.28523453 -0.2250839 -0.1540565 -0.3335620 -0.2311809
#>  [2,] -0.06872328 -0.08974227 -0.1732156 -0.2787165 -0.1106945 -0.2609753
#>  [3,] -0.08623735 -0.11029672 -0.2097704 -0.3646658 -0.2075587 -0.3047873
#>  [4,] -0.17645095 -0.04169105 -0.2152172 -0.2506967 -0.2427076 -0.3066257
#>  [5,] -0.25399433 -0.24128215 -0.2716077 -0.2708324 -0.1697890 -0.2781767
#>  [6,] -0.15258037 -0.25084413 -0.2706897 -0.1825810 -0.1380871 -0.3667360
#>  [7,] -0.18740755 -0.20810040 -0.2561417 -0.1985388 -0.3434876 -0.4303770
#>  [8,] -0.23198001 -0.20938990 -0.2228535 -0.2378855 -0.3724387 -0.3237573
#>  [9,] -0.16268240 -0.20885573 -0.2389394 -0.3573050 -0.1891862 -0.3950633
#> [10,] -0.14357693 -0.14093813 -0.1775962 -0.3321246 -0.3127433 -0.3079530
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3817474 -0.2489295 -0.3530033 -0.1844349 -0.3552726 -0.3740392
#>  [2,] -0.2947665 -0.1839243 -0.2642753 -0.2480591 -0.2511839 -0.3126911
#>  [3,] -0.3844462 -0.3102808 -0.3448090 -0.4224591 -0.4054218 -0.3506977
#>  [4,] -0.3571344 -0.2287143 -0.3654608 -0.2566314 -0.3764111 -0.3249879
#>  [5,] -0.2225175 -0.3610892 -0.2676479 -0.3811986 -0.3876640 -0.2652658
#>  [6,] -0.2700123 -0.2479375 -0.2167346 -0.3606085 -0.3778247 -0.3611416
#>  [7,] -0.4654713 -0.3695818 -0.4689616 -0.3962005 -0.4425157 -0.3996580
#>  [8,] -0.4153608 -0.5110802 -0.4635266 -0.4322026 -0.4209039 -0.3413526
#>  [9,] -0.2668869 -0.2978674 -0.2916270 -0.3812847 -0.2562045 -0.3033724
#> [10,] -0.2134297 -0.2714089 -0.2103970 -0.2228595 -0.2539738 -0.3958889
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.3414991 -0.3709891 -0.2578077 -0.15509232 -0.2836638 -0.26335636
#>  [2,] -0.2998045 -0.2067820 -0.1675764 -0.09295205 -0.3422993 -0.13499410
#>  [3,] -0.3326882 -0.2463130 -0.3210643 -0.25907633 -0.3373016 -0.25372278
#>  [4,] -0.3898463 -0.3216945 -0.3715826 -0.20675070 -0.2788488 -0.05377174
#>  [5,] -0.3893776 -0.3025906 -0.2795951 -0.09393546 -0.2515232 -0.14357879
#>  [6,] -0.4072611 -0.3154797 -0.2323325 -0.25865883 -0.3305989 -0.19978210
#>  [7,] -0.4019705 -0.3806405 -0.4826850 -0.25738234 -0.4450560 -0.24660275
#>  [8,] -0.3663827 -0.3599337 -0.3606490 -0.23542692 -0.4723217 -0.17967030
#>  [9,] -0.2669944 -0.4008561 -0.2910304 -0.23863681 -0.3465059 -0.14505447
#> [10,] -0.2637450 -0.3483352 -0.3908327 -0.09055233 -0.3260699 -0.26619210
#>                V8          V9
#>  [1,] -0.10752424 -0.19200554
#>  [2,] -0.13109882 -0.13515769
#>  [3,] -0.18380832 -0.13237063
#>  [4,] -0.12445513 -0.01784274
#>  [5,] -0.15688486 -0.08092049
#>  [6,] -0.23155427 -0.05697472
#>  [7,] -0.09341656 -0.21717532
#>  [8,] -0.15656702 -0.09829025
#>  [9,] -0.22149724 -0.10174514
#> [10,] -0.14765083 -0.21539543
#> 
#> $W[[2]]
#>              [,1]        [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,]  0.029967571 -0.04785002 0.02089894 0.07231749  0.01341678 0.08089101
#> [2,] -0.003509746 -0.03982994 0.08776222 0.05210436 -0.06881765 0.05912672
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.05316714  0.08510703 -0.07562717  0.01221892
#> [2,] -0.10542510 -0.13118518  0.05946136 -0.02236064
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -4.595478e-05 -5.108318e-04 -5.978728e-04 -0.0006076177 -0.0003830811
#>  [2,]  5.565802e-05  4.457898e-04  5.249215e-04  0.0005509183  0.0005318741
#>  [3,]  1.393005e-04  1.218835e-03  1.371727e-03  0.0013277993  0.0012035769
#>  [4,]  6.748124e-05  4.735900e-04  5.708953e-04  0.0006115590  0.0005816346
#>  [5,] -5.985702e-05 -7.116413e-04 -7.457108e-04 -0.0006269858 -0.0004047033
#>  [6,]  6.701942e-05  4.905120e-04  5.965523e-04  0.0006506312  0.0006156968
#>  [7,]  1.545111e-05  8.627790e-06  7.087576e-05  0.0001707282  0.0002060069
#>  [8,] -1.334022e-04 -1.308383e-03 -1.316619e-03 -0.0010254297 -0.0008105031
#>  [9,]  2.269552e-04  2.016007e-03  2.233245e-03  0.0020917216  0.0018644070
#> [10,] -4.086486e-05 -4.573259e-04 -5.204773e-04 -0.0005089230 -0.0003125622
#>                 V14           V15           V16           V17           V18
#>  [1,] -6.006586e-05  0.0001755429  0.0002059301  1.422604e-04  1.884276e-05
#>  [2,]  4.581624e-04  0.0004223313  0.0004663443  5.305023e-04  5.822512e-04
#>  [3,]  9.420032e-04  0.0007173258  0.0007551379  8.695256e-04  1.043965e-03
#>  [4,]  4.953772e-04  0.0004893591  0.0005529070  6.186221e-04  6.345058e-04
#>  [5,] -7.966839e-05  0.0002672337  0.0003565781  3.331318e-04  1.332892e-04
#>  [6,]  5.136420e-04  0.0004876482  0.0005476905  6.212422e-04  6.482781e-04
#>  [7,]  2.198543e-04  0.0003155455  0.0003924482  4.440371e-04  4.044065e-04
#>  [8,] -5.404700e-04 -0.0002053552 -0.0001193969 -9.894761e-05 -2.628570e-04
#>  [9,]  1.429654e-03  0.0009922664  0.0009883514  1.136939e-03  1.458497e-03
#> [10,] -2.306294e-05  0.0002221103  0.0002739931  2.351843e-04  9.692546e-05
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0002219467 -5.777680e-05 -0.0003499964 -0.0002800729 -1.185960e-04
#>  [2,]  0.0007311573  7.767633e-05  0.0008613861  0.0008982601  8.620911e-04
#>  [3,]  0.0015128809  1.860493e-04  0.0019465326  0.0019636306  1.721704e-03
#>  [4,]  0.0007842690  9.499576e-05  0.0009202564  0.0009537899  8.913681e-04
#>  [5,] -0.0002619078 -6.861057e-05 -0.0005481319 -0.0004436463 -1.507245e-04
#>  [6,]  0.0007987949  9.334204e-05  0.0009327015  0.0009780844  9.489307e-04
#>  [7,]  0.0003806392  2.963441e-05  0.0003150786  0.0003720592  4.656227e-04
#>  [8,] -0.0007207598 -1.619754e-04 -0.0013497293 -0.0012356638 -7.110864e-04
#>  [9,]  0.0022666847  2.947292e-04  0.0030491313  0.0030386019  2.548390e-03
#> [10,] -0.0001567438 -4.860942e-05 -0.0002890875 -0.0002062802 -2.406216e-05
#>                V23          V24          V25          V26          V27
#>  [1,] 2.565884e-04 0.0004594586 0.0005420044 0.0005744007 0.0004328009
#>  [2,] 8.065337e-04 0.0008125215 0.0007922407 0.0008347168 0.0008878654
#>  [3,] 1.325776e-03 0.0011850884 0.0011237456 0.0011750854 0.0013104387
#>  [4,] 8.421277e-04 0.0008519952 0.0008232651 0.0008700437 0.0009440267
#>  [5,] 3.991108e-04 0.0006955506 0.0007747097 0.0008134697 0.0007013311
#>  [6,] 9.364159e-04 0.0009672981 0.0009523472 0.0010157499 0.0011061877
#>  [7,] 6.368271e-04 0.0007476823 0.0007339769 0.0007805515 0.0008066823
#>  [8,] 8.434093e-05 0.0004285523 0.0004293343 0.0004791619 0.0003802134
#>  [9,] 1.790741e-03 0.0014755469 0.0013551314 0.0013983606 0.0015997919
#> [10,] 3.379318e-04 0.0005449973 0.0006227643 0.0006489761 0.0005263388
#>                 V28          V29            V3          V30          V31
#>  [1,]  1.910362e-04 0.0002401717 -5.070210e-05 0.0004316064 0.0005029190
#>  [2,]  9.131685e-04 0.0008368756  8.200709e-05 0.0007208144 0.0005605884
#>  [3,]  1.529179e-03 0.0013440717  1.903065e-04 0.0009970300 0.0006430182
#>  [4,]  9.869184e-04 0.0009285092  9.523306e-05 0.0008492521 0.0006756054
#>  [5,]  3.654933e-04 0.0004661701 -6.288321e-05 0.0007622291 0.0008454205
#>  [6,]  1.137281e-03 0.0010556055  9.408029e-05 0.0009457970 0.0007683919
#>  [7,]  7.186128e-04 0.0007216318  3.142745e-05 0.0007676632 0.0007068741
#>  [8,] -5.645739e-05 0.0001212913 -1.621855e-04 0.0004576208 0.0006977072
#>  [9,]  2.042664e-03 0.0017578511  2.989497e-04 0.0011572029 0.0006018784
#> [10,]  2.840239e-04 0.0003297936 -4.166751e-05 0.0005340921 0.0005798146
#>                V32          V33          V34          V35           V36
#>  [1,] 2.227297e-04 2.805677e-04 0.0005404083 0.0006888257  7.312703e-04
#>  [2,] 5.163322e-04 4.392086e-04 0.0003394209 0.0002719353  2.584016e-04
#>  [3,] 8.121922e-04 7.077248e-04 0.0004179972 0.0002126707  9.922683e-05
#>  [4,] 6.053998e-04 5.175106e-04 0.0004267244 0.0003421585  2.986465e-04
#>  [5,] 3.760110e-04 3.456386e-04 0.0005966774 0.0007442922  8.572868e-04
#>  [6,] 6.758359e-04 5.622364e-04 0.0004520930 0.0003481171  3.156854e-04
#>  [7,] 4.917212e-04 4.086345e-04 0.0004369145 0.0004501427  4.906144e-04
#>  [8,] 7.445084e-05 2.653237e-05 0.0004400534 0.0007417591  9.943750e-04
#>  [9,] 1.022664e-03 9.354725e-04 0.0004687317 0.0001279749 -1.134566e-04
#> [10,] 2.801928e-04 2.967576e-04 0.0004966327 0.0006118908  6.752181e-04
#>                V37          V38          V39            V4          V40
#>  [1,] 0.0004839984 0.0001614223 5.725728e-05 -9.984573e-05 0.0001951210
#>  [2,] 0.0003088646 0.0003678401 3.914280e-04  1.122048e-04 0.0003393669
#>  [3,] 0.0002450078 0.0005625651 7.010797e-04  2.905102e-04 0.0004845389
#>  [4,] 0.0003579432 0.0004304580 4.372182e-04  1.284419e-04 0.0003747908
#>  [5,] 0.0006834466 0.0002531848 5.170635e-05 -1.375520e-04 0.0002696203
#>  [6,] 0.0003916217 0.0004576336 4.697168e-04  1.282804e-04 0.0004136620
#>  [7,] 0.0004723351 0.0003796024 3.239206e-04  1.571329e-05 0.0003649657
#>  [8,] 0.0007291642 0.0002315194 5.468438e-05 -3.068144e-04 0.0003635708
#>  [9,] 0.0001046911 0.0007168748 9.963701e-04  4.770277e-04 0.0005902310
#> [10,] 0.0004893916 0.0001629987 3.780307e-05 -8.978398e-05 0.0001936664
#>                V41           V42           V43           V44           V45
#>  [1,] 0.0001450615 -7.717613e-06 -9.481071e-05 -0.0002533672 -0.0003753165
#>  [2,] 0.0003809255  4.279034e-04  4.108108e-04  0.0003950302  0.0004161426
#>  [3,] 0.0005598637  7.301259e-04  7.551483e-04  0.0008505247  0.0010217151
#>  [4,] 0.0004278938  4.836430e-04  4.757393e-04  0.0004557661  0.0004886521
#>  [5,] 0.0002475191  8.868413e-05 -1.095988e-05 -0.0002458344 -0.0004488064
#>  [6,] 0.0004628639  5.211518e-04  5.157459e-04  0.0004935145  0.0005194824
#>  [7,] 0.0004044790  3.757251e-04  3.294372e-04  0.0002329534  0.0001497465
#>  [8,] 0.0003261489  9.082003e-05 -1.605510e-05 -0.0003412356 -0.0008108285
#>  [9,] 0.0006680217  9.746416e-04  1.047293e-03  0.0012448686  0.0015712278
#> [10,] 0.0001709912  3.259895e-05 -5.621807e-05 -0.0002062428 -0.0003112591
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0002442350 -0.0001738467 -1.639950e-04 -1.026376e-04 -7.684183e-05
#>  [2,]  0.0003112896  0.0002291045  1.825066e-04  1.051092e-04  1.399790e-04
#>  [3,]  0.0007293216  0.0005070682  4.331123e-04  2.630263e-04  2.953847e-04
#>  [4,]  0.0003709105  0.0002650753  2.019085e-04  1.139979e-04  1.629689e-04
#>  [5,] -0.0002632922 -0.0001625297 -1.848502e-04 -1.309274e-04 -6.610436e-05
#>  [6,]  0.0003961547  0.0002872651  2.169015e-04  1.211287e-04  1.681825e-04
#>  [7,]  0.0001354601  0.0001149179  6.178982e-05  2.603496e-05  7.344583e-05
#>  [8,] -0.0005628699 -0.0003468191 -3.641005e-04 -2.317278e-04 -1.654197e-04
#>  [9,]  0.0011010802  0.0007457505  6.646775e-04  4.134213e-04  4.476054e-04
#> [10,] -0.0001846988 -0.0001231563 -1.294491e-04 -8.684960e-05 -5.787183e-05
#>                 V50           V51           V52           V53           V54
#>  [1,] -1.752566e-05 -1.546865e-05 -1.592087e-05 -2.797747e-06 -6.446620e-06
#>  [2,]  3.283227e-05  2.906435e-05  2.775782e-05  1.609292e-05  1.827510e-05
#>  [3,]  7.894981e-05  6.551276e-05  6.396425e-05  3.440365e-05  4.287722e-05
#>  [4,]  3.726121e-05  3.287019e-05  3.145713e-05  1.895446e-05  1.870326e-05
#>  [5,] -2.759174e-05 -1.924525e-05 -2.106033e-05 -5.926589e-06 -1.393523e-05
#>  [6,]  3.872445e-05  3.440720e-05  3.201281e-05  1.908539e-05  1.923550e-05
#>  [7,]  1.089966e-05  1.280022e-05  1.064714e-05  8.324156e-06  4.388639e-06
#>  [8,] -6.658459e-05 -4.810591e-05 -5.192678e-05 -2.172835e-05 -3.833436e-05
#>  [9,]  1.240376e-04  1.005614e-04  9.966469e-05  5.271536e-05  6.871164e-05
#> [10,] -1.520577e-05 -1.294947e-05 -1.356103e-05 -3.261438e-06 -6.112482e-06
#>                 V55           V56           V57           V58           V59
#>  [1,] -2.628806e-06 -1.789144e-06  5.080850e-06 -2.817344e-06 -4.802699e-06
#>  [2,]  1.334989e-05  1.188425e-05  9.107856e-06  1.321287e-05  1.150778e-05
#>  [3,]  3.346162e-05  2.499570e-05  1.590596e-05  2.825947e-05  2.722553e-05
#>  [4,]  1.482142e-05  1.285685e-05  1.087272e-05  1.461197e-05  1.230671e-05
#>  [5,] -1.133923e-05 -3.745716e-06  5.666963e-06 -5.832503e-06 -9.690378e-06
#>  [6,]  1.418699e-05  1.310268e-05  1.091238e-05  1.495903e-05  1.210355e-05
#>  [7,]  2.981753e-06  5.020881e-06  7.069715e-06  5.794540e-06  3.755855e-06
#>  [8,] -3.003222e-05 -1.756059e-05 -4.931262e-06 -1.804195e-05 -2.069909e-05
#>  [9,]  5.563614e-05  3.888698e-05  2.308577e-05  4.523034e-05  4.405727e-05
#> [10,] -4.464340e-06 -1.535329e-06  5.439294e-06 -3.607267e-06 -5.473633e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -4.522796e-05  1.040823e-06 -6.267392e-05 -1.422937e-04 -3.453676e-04
#>  [2,]  1.669404e-04  7.227618e-06  1.681961e-04  2.306523e-04  3.736959e-04
#>  [3,]  3.146103e-04  1.371845e-05  3.473734e-04  5.328209e-04  9.715515e-04
#>  [4,]  1.884100e-04  8.436150e-06  1.910776e-04  2.499770e-04  4.003135e-04
#>  [5,] -1.186290e-05  7.447770e-07 -5.625416e-05 -1.844012e-04 -4.882802e-04
#>  [6,]  2.028988e-04  8.641544e-06  2.030584e-04  2.624461e-04  4.158928e-04
#>  [7,]  1.102541e-04  4.901448e-06  9.020180e-05  7.414324e-05  4.227318e-05
#>  [8,] -1.082492e-04 -5.029451e-06 -1.901038e-04 -4.278414e-04 -9.831985e-04
#>  [9,]  4.548892e-04  2.079977e-05  5.232422e-04  8.426114e-04  1.582896e-03
#> [10,] -2.239080e-05  6.861451e-07 -4.427586e-05 -1.145156e-04 -2.983249e-04
#> 
#> $vW[[2]]
#>             [,1]         [,2]         [,3]         [,4]          [,5]
#> [1,]  0.01106201  0.001121186  0.004429678 -0.007357516 -0.0003749659
#> [2,] -0.01106201 -0.001121186 -0.004429678  0.007357516  0.0003749659
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] -0.007461445 -0.003826953 -0.009127519  0.007051766  0.007731437
#> [2,]  0.007461445  0.003826953  0.009127519 -0.007051766 -0.007731437
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.01207836 0.07640003 0.41858889 0.56049183 0.30075086 0.34632273
#>  [7] 0.52085042 0.60086209 0.43388280 0.24395149
#> 
#> $B[[2]]
#> [1]  0.09238048 -0.08433691
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0006577655 0.0012362610 0.0019375972 0.0013634372 0.0008551807
#>  [6] 0.0014751892 0.0010844640 0.0002409256 0.0025335291 0.0007073077
#> 
#> $vB[[2]]
#> [1] -0.01664614  0.01664614
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#>  [2,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#>  [3,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#>  [4,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#>  [5,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#>  [6,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#>  [7,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#>  [8,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [9,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [10,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [11,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [12,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [13,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [14,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [15,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [16,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [17,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [18,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [19,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [20,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [21,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [22,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [23,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [24,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [25,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [26,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [27,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [28,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [29,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [30,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [31,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [32,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [33,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [34,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [35,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [36,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [37,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [38,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#> [39,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [40,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#>  [2,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#>  [3,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#>  [4,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#>  [5,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#>  [6,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#>  [7,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#>  [8,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [9,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [10,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [11,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [12,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [13,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [14,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [15,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [16,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [17,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [18,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [19,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [20,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [21,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [22,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [23,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [24,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [25,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [26,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [27,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [28,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [29,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [30,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [31,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [32,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [33,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [34,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [35,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [36,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [37,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [38,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#> [39,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [40,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#>  [2,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#>  [3,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#>  [4,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#>  [5,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#>  [6,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#>  [7,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#>  [8,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [9,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [10,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [11,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [12,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [13,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [14,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [15,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [16,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [17,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [18,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [19,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [20,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [21,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [22,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [23,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [24,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [25,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [26,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [27,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [28,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [29,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [30,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [31,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [32,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [33,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [34,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [35,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [36,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [37,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [38,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#> [39,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [40,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#>  [2,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#>  [3,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#>  [4,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#>  [5,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#>  [6,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#>  [7,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#>  [8,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [9,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [10,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [11,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [12,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [13,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [14,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [15,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [16,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [17,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [18,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [19,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [20,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [21,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [22,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [23,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [24,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [25,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [26,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [27,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [28,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [29,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [30,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [31,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [32,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [33,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [34,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [35,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [36,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [37,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [38,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#> [39,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [40,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#>  [2,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#>  [3,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#>  [4,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#>  [5,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#>  [6,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#>  [7,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#>  [8,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [9,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [10,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [11,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [12,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [13,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [14,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [15,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [16,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [17,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [18,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [19,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [20,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [21,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [22,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [23,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [24,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [25,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [26,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [27,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [28,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [29,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [30,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [31,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [32,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [33,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [34,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [35,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [36,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [37,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [38,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#> [39,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [40,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#>  [2,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#>  [3,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#>  [4,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#>  [5,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#>  [6,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#>  [7,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#>  [8,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [9,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [10,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [11,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [12,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [13,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [14,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [15,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [16,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [17,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [18,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [19,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [20,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [21,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [22,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [23,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [24,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [25,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [26,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [27,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [28,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [29,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [30,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [31,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [32,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [33,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [34,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [35,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [36,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [37,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [38,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#> [39,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [40,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1796520 0.4536469 0.5599345 0.6871856 0.5036775 0.7204989 0.6464338
#>  [2,] 0.1460012 0.4906214 0.5776922 0.7698798 0.4365626 0.7501414 0.7834140
#>  [3,] 0.2958816 0.5267600 0.6814545 0.8024959 0.5483755 0.7205917 0.7860370
#>  [4,] 0.2103416 0.4333558 0.5899378 0.7303788 0.4379620 0.6006214 0.7208332
#>  [5,] 0.3148827 0.5526669 0.7376415 0.8333456 0.5910082 0.7362651 0.8422637
#>  [6,] 0.1738213 0.4558409 0.6345964 0.8035537 0.5841210 0.7515672 0.7676993
#>  [7,] 0.2023323 0.5315095 0.6744861 0.8238194 0.5450927 0.8008929 0.8393638
#>  [8,] 0.2690304 0.5814673 0.7735426 0.8722144 0.6944845 0.8386800 0.8746776
#>  [9,] 0.2075380 0.4976414 0.6360335 0.8134851 0.5252007 0.7100312 0.8083726
#> [10,] 0.2335215 0.5345748 0.6975349 0.8410714 0.6065100 0.7928073 0.8315931
#> [11,] 0.2627554 0.5570377 0.7501844 0.8609729 0.6243643 0.7796010 0.8678492
#> [12,] 0.2370154 0.4931758 0.6740848 0.8126965 0.5095213 0.7390942 0.8162855
#> [13,] 0.3122264 0.5398836 0.7200737 0.8144159 0.5579457 0.6852645 0.8209486
#> [14,] 0.3120227 0.5914162 0.7925180 0.8793272 0.7036562 0.8224883 0.8806100
#> [15,] 0.1695086 0.5256252 0.6971350 0.8314142 0.4597517 0.7515456 0.8176975
#> [16,] 0.3166123 0.5560149 0.7487141 0.8433189 0.5554049 0.7109204 0.8560505
#> [17,] 0.1539827 0.4095568 0.4490405 0.6378552 0.3572070 0.6303784 0.5696209
#> [18,] 0.1327364 0.4087170 0.4688721 0.6690333 0.3094397 0.6211154 0.6460334
#> [19,] 0.3876177 0.5348503 0.7130316 0.7967816 0.5899927 0.6666555 0.8006421
#> [20,] 0.2127377 0.5211428 0.6462950 0.8086827 0.5834829 0.8057899 0.7962342
#> [21,] 0.1112939 0.4017839 0.4315425 0.6449566 0.3182491 0.6674743 0.5866089
#> [22,] 0.3424305 0.5619921 0.7542661 0.8428402 0.5901785 0.6931242 0.8577555
#> [23,] 0.3289573 0.5641463 0.7486003 0.8551367 0.6334600 0.7557311 0.8596278
#> [24,] 0.1210445 0.4257823 0.4885712 0.6733782 0.3311109 0.6704356 0.6401651
#> [25,] 0.2466820 0.5420061 0.7004067 0.8550666 0.5333799 0.7587639 0.8345458
#> [26,] 0.3027850 0.5810475 0.7549880 0.8525302 0.6863093 0.8201530 0.8661241
#> [27,] 0.3183756 0.5610024 0.7546602 0.8600242 0.6393064 0.7591436 0.8626804
#> [28,] 0.2885884 0.5162757 0.6907176 0.7954364 0.5283941 0.6692250 0.8063453
#> [29,] 0.3319113 0.5268070 0.6678084 0.8020576 0.5934762 0.6871982 0.7829119
#> [30,] 0.2149401 0.5050354 0.6239884 0.8252758 0.5242302 0.7423874 0.7912074
#> [31,] 0.3096013 0.5595061 0.7476073 0.8482976 0.5965235 0.7437306 0.8553361
#> [32,] 0.1820876 0.4456430 0.5895772 0.7058639 0.5128171 0.6987439 0.6627833
#> [33,] 0.1688471 0.5031692 0.6918555 0.8104021 0.5734602 0.7787790 0.8158264
#> [34,] 0.1578135 0.4373228 0.5859018 0.7461368 0.3583612 0.6393574 0.7553628
#> [35,] 0.2032719 0.4705412 0.5791716 0.7951522 0.5162410 0.6733295 0.7566198
#> [36,] 0.2545732 0.5354605 0.6879717 0.8418190 0.5271451 0.7481583 0.8199498
#> [37,] 0.3432733 0.5846703 0.7645784 0.8637392 0.6111060 0.7701780 0.8673277
#> [38,] 0.2546898 0.5187917 0.6920586 0.8106858 0.5193200 0.7030327 0.8210071
#> [39,] 0.1988929 0.4514389 0.5358236 0.6844838 0.3449446 0.6074237 0.6478567
#> [40,] 0.2643589 0.5132033 0.7437464 0.8160034 0.5775908 0.7254724 0.8359532
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7829821 0.4866427 0.3210822
#>  [2,] 0.8280683 0.6148565 0.2060225
#>  [3,] 0.8352106 0.7209734 0.3926862
#>  [4,] 0.7701903 0.6246037 0.3177396
#>  [5,] 0.8646387 0.7701768 0.4492255
#>  [6,] 0.8785973 0.5966973 0.3718188
#>  [7,] 0.8764868 0.6883247 0.3034348
#>  [8,] 0.9261694 0.7544351 0.4920481
#>  [9,] 0.8608955 0.6733447 0.3203952
#> [10,] 0.9076213 0.6992869 0.4038741
#> [11,] 0.9083069 0.7655265 0.4386031
#> [12,] 0.8593488 0.7089449 0.3410931
#> [13,] 0.8332346 0.7617719 0.4300708
#> [14,] 0.9285617 0.7833042 0.5408016
#> [15,] 0.8809634 0.7461416 0.2672301
#> [16,] 0.8714515 0.8016630 0.4114288
#> [17,] 0.6906897 0.4828014 0.2074593
#> [18,] 0.7334803 0.5399801 0.1650085
#> [19,] 0.8231361 0.7456001 0.5036952
#> [20,] 0.8912156 0.6422319 0.3573690
#> [21,] 0.7070463 0.4860126 0.1588735
#> [22,] 0.8531661 0.7910954 0.4715098
#> [23,] 0.8930800 0.7655381 0.4935822
#> [24,] 0.7406322 0.5461193 0.1598547
#> [25,] 0.9032642 0.7533607 0.3665786
#> [26,] 0.9088212 0.7420920 0.4993183
#> [27,] 0.9000097 0.7631777 0.4982019
#> [28,] 0.8206995 0.7353431 0.3960077
#> [29,] 0.8423211 0.6992835 0.4604344
#> [30,] 0.8615080 0.6657439 0.3241122
#> [31,] 0.8775741 0.7846348 0.4476824
#> [32,] 0.7757447 0.5245059 0.3470662
#> [33,] 0.8834207 0.6568433 0.3635666
#> [34,] 0.7739434 0.6409656 0.2121273
#> [35,] 0.8442766 0.6180015 0.3194943
#> [36,] 0.8905016 0.7441052 0.3648254
#> [37,] 0.8903220 0.8072468 0.4635887
#> [38,] 0.8497384 0.7435121 0.3526894
#> [39,] 0.7070264 0.6173275 0.2124960
#> [40,] 0.8770777 0.7389291 0.4303771
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5766774 0.4233226
#>  [2,] 0.5732176 0.4267824
#>  [3,] 0.5754737 0.4245263
#>  [4,] 0.5713156 0.4286844
#>  [5,] 0.5770879 0.4229121
#>  [6,] 0.5806534 0.4193466
#>  [7,] 0.5768438 0.4231562
#>  [8,] 0.5825029 0.4174971
#>  [9,] 0.5765038 0.4234962
#> [10,] 0.5804870 0.4195130
#> [11,] 0.5793646 0.4206354
#> [12,] 0.5753192 0.4246808
#> [13,] 0.5748132 0.4251868
#> [14,] 0.5828904 0.4171096
#> [15,] 0.5721403 0.4278597
#> [16,] 0.5752999 0.4247001
#> [17,] 0.5682367 0.4317633
#> [18,] 0.5674341 0.4325659
#> [19,] 0.5771948 0.4228052
#> [20,] 0.5804086 0.4195914
#> [21,] 0.5674960 0.4325040
#> [22,] 0.5764059 0.4235941
#> [23,] 0.5802527 0.4197473
#> [24,] 0.5675297 0.4324703
#> [25,] 0.5768952 0.4231048
#> [26,] 0.5825259 0.4174741
#> [27,] 0.5806497 0.4193503
#> [28,] 0.5738994 0.4261006
#> [29,] 0.5787737 0.4212263
#> [30,] 0.5769276 0.4230724
#> [31,] 0.5773294 0.4226706
#> [32,] 0.5754994 0.4245006
#> [33,] 0.5784350 0.4215650
#> [34,] 0.5679264 0.4320736
#> [35,] 0.5772817 0.4227183
#> [36,] 0.5765041 0.4234959
#> [37,] 0.5781384 0.4218616
#> [38,] 0.5741065 0.4258935
#> [39,] 0.5650500 0.4349500
#> [40,] 0.5772883 0.4227117
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]         [,2]        [,3]      [,4]        [,5]      [,6]
#>  [1,] -1.5187073 -0.185946202  0.24089643 0.7869943  0.01471011 0.9469378
#>  [2,] -1.7663149 -0.037518863  0.31330661 1.2076323 -0.25512447 1.0993668
#>  [3,] -0.8669873  0.107142502  0.76046411 1.4019676  0.19410926 0.9473985
#>  [4,] -1.3228673 -0.268172383  0.36370812 0.9965451 -0.24943735 0.4080548
#>  [5,] -0.7773899  0.211452124  1.03374589 1.6095260  0.36813499 1.0266459
#>  [6,] -1.5587831 -0.177097655  0.55198661 1.4086548  0.33971381 1.1069883
#>  [7,] -1.3717809  0.126205415  0.72854608 1.5424420  0.18086226 1.3918841
#>  [8,] -0.9995474  0.328799705  1.22842401 1.9206814  0.82116933 1.6484393
#>  [9,] -1.3398302 -0.009434507  0.55818926 1.4728163  0.10088841 0.8955356
#> [10,] -1.1885326  0.138520371  0.83558649 1.6662213  0.43266545 1.3419311
#> [11,] -1.0316964  0.229148093  1.09959615 1.8233940  0.50811413 1.2633429
#> [12,] -1.1691125 -0.027298490  0.72671853 1.4676272  0.03808991 1.0412662
#> [13,] -0.7897314  0.159874026  0.94482711 1.4789629  0.23282898 0.7780723
#> [14,] -0.7906800  0.369822888  1.34017091 1.9860744  0.86476940 1.5332979
#> [15,] -1.5891140  0.102590724  0.83369214 1.5956834 -0.16134239 1.1068726
#> [16,] -0.7693845  0.225003958  1.09176561 1.6831324  0.22253329 0.8998583
#> [17,] -1.7036993 -0.365797679 -0.20454832 0.5660671 -0.58750733 0.5338404
#> [18,] -1.8769779 -0.369271810 -0.12467288 0.7038163 -0.80274027 0.4942851
#> [19,] -0.4573371  0.139627454  0.91015382 1.3662993  0.36393507 0.6930970
#> [20,] -1.3085013  0.084621572  0.60279298 1.4414732  0.33708770 1.4228824
#> [21,] -2.0775917 -0.398037550 -0.27556055 0.5969429 -0.76183002 0.6967839
#> [22,] -0.6524815  0.249250970  1.12149624 1.6795145  0.36470330 0.8147661
#> [23,] -0.7129048  0.258007086  1.09116116 1.7754705  0.54708923 1.1294157
#> [24,] -1.9825763 -0.299080553 -0.04572313 0.7235044 -0.70316503 0.7101559
#> [25,] -1.1163875  0.168421304  0.84923507 1.7749052  0.13371848 1.1459147
#> [26,] -0.8340708  0.327074853  1.12539480 1.7545852  0.78292102 1.5173841
#> [27,] -0.7612472  0.245231172  1.12362365 1.8154911  0.57235513 1.1479900
#> [28,] -0.9022498  0.065125905  0.80347609 1.3580118  0.11369882 0.7046821
#> [29,] -0.6995531  0.107331086  0.69828958 1.3992046  0.37835460 0.7870529
#> [30,] -1.2954007  0.020142179  0.50651191 1.5525090  0.09699683 1.0584141
#> [31,] -0.8019837  0.239157637  1.08589186 1.7213108  0.39100032 1.0654497
#> [32,] -1.5022671 -0.218290666  0.36221768 0.8753800  0.05127972 0.8413235
#> [33,] -1.5938202  0.012676803  0.80880815 1.4526250  0.29598303 1.2585654
#> [34,] -1.6745873 -0.252034515  0.34704904 1.0781132 -0.58248393 0.5725762
#> [35,] -1.3659691 -0.117971726  0.31937363 1.3562664  0.06498677 0.7232830
#> [36,] -1.0743684  0.142080478  0.79065375 1.6718249  0.10868720 1.0888137
#> [37,] -0.6487409  0.341975389  1.17794659 1.8467002  0.45196356 1.2093165
#> [38,] -1.0737540  0.075202151  0.80976110 1.4544727  0.07731845 0.8617814
#> [39,] -1.3932278 -0.194858586  0.14354052 0.7744547 -0.64133612 0.4364955
#> [40,] -1.0234345  0.052825528  1.06553229 1.4895015  0.31289116 0.9717710
#>            [,7]      [,8]        [,9]        [,10]
#>  [1,] 0.6034002 1.2831301 -0.05344190 -0.748802715
#>  [2,] 1.2856735 1.5719982  0.46777278 -1.349069668
#>  [3,] 1.3012008 1.6230158  0.94929514 -0.436034726
#>  [4,] 0.9485985 1.2093860  0.50913500 -0.764179108
#>  [5,] 1.6751684 1.8543642  1.20930952 -0.203800332
#>  [6,] 1.1953652 1.9792135  0.39172254 -0.524421936
#>  [7,] 1.6535020 1.9595741  0.79229890 -0.830994441
#>  [8,] 1.9429656 2.5292834  1.12240809 -0.031810282
#>  [9,] 1.4394703 1.8227480  0.72335191 -0.751956302
#> [10,] 1.5969600 2.2849310  0.84390447 -0.389348647
#> [11,] 1.8820741 2.2931348  1.18322114 -0.246833429
#> [12,] 1.4913811 1.8098915  0.89026504 -0.658426993
#> [13,] 1.5227879 1.6087273  1.16241872 -0.281562258
#> [14,] 1.9982198 2.5648030  1.28502653  0.163569972
#> [15,] 1.5008252 2.0015848  1.07813856 -1.008721675
#> [16,] 1.7828664 1.9138541  1.39672077 -0.358061745
#> [17,] 0.2803045 0.8033458 -0.06882150 -1.340308780
#> [18,] 0.6016486 1.0123527  0.16026266 -1.621424591
#> [19,] 1.3903122 1.5377410  1.07528181  0.014780940
#> [20,] 1.3629219 2.1032185  0.58506449 -0.586801749
#> [21,] 0.3499644 0.8810815 -0.05596405 -1.666633898
#> [22,] 1.7967715 1.7596525  1.33154083 -0.114084409
#> [23,] 1.8122023 2.1225957  1.18328578 -0.025672504
#> [24,] 0.5760808 1.0492568  0.18500313 -1.659309772
#> [25,] 1.6181934 2.2340318  1.11661741 -0.546922962
#> [26,] 1.8671152 2.2993264  1.05687033 -0.002726986
#> [27,] 1.8377333 2.1973323  1.17018060 -0.007192500
#> [28,] 1.4264355 1.5210939  1.02190288 -0.422127628
#> [29,] 1.2827169 1.6756004  0.84388822 -0.158593925
#> [30,] 1.3322185 1.8278718  0.68899768 -0.734937428
#> [31,] 1.7770817 1.9696557  1.29288308 -0.210039358
#> [32,] 0.6757222 1.2410384  0.09810217 -0.631960260
#> [33,] 1.4883226 2.0252296  0.64925845 -0.559917178
#> [34,] 1.1274221 1.2307132  0.57955745 -1.312149810
#> [35,] 1.1342361 1.6903990  0.48107416 -0.756096774
#> [36,] 1.5160076 2.0958750  1.06741578 -0.554481084
#> [37,] 1.8775346 2.0940341  1.43221874 -0.145903612
#> [38,] 1.5231855 1.7325506  1.06430374 -0.607238666
#> [39,] 0.6096314 0.8809857  0.47821999 -1.309945331
#> [40,] 1.6284209 1.9650432  1.04040984 -0.280312867
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1758854 -0.1332629
#>  [2,] 0.1644306 -0.1305604
#>  [3,] 0.1697669 -0.1344524
#>  [4,] 0.1572592 -0.1299617
#>  [5,] 0.1712552 -0.1395750
#>  [6,] 0.1821430 -0.1433133
#>  [7,] 0.1733242 -0.1365061
#>  [8,] 0.1850936 -0.1479630
#>  [9,] 0.1684911 -0.1399465
#> [10,] 0.1804329 -0.1443401
#> [11,] 0.1766083 -0.1435573
#> [12,] 0.1701545 -0.1334329
#> [13,] 0.1649157 -0.1366009
#> [14,] 0.1851026 -0.1495474
#> [15,] 0.1654707 -0.1251183
#> [16,] 0.1665912 -0.1369171
#> [17,] 0.1577136 -0.1169470
#> [18,] 0.1522338 -0.1191561
#> [19,] 0.1689397 -0.1423287
#> [20,] 0.1812125 -0.1432385
#> [21,] 0.1576443 -0.1139980
#> [22,] 0.1670643 -0.1409721
#> [23,] 0.1773467 -0.1464641
#> [24,] 0.1559960 -0.1157836
#> [25,] 0.1732024 -0.1368383
#> [26,] 0.1834433 -0.1497077
#> [27,] 0.1784638 -0.1469771
#> [28,] 0.1625221 -0.1352565
#> [29,] 0.1725624 -0.1451789
#> [30,] 0.1728936 -0.1372798
#> [31,] 0.1717782 -0.1400418
#> [32,] 0.1732854 -0.1310392
#> [33,] 0.1768159 -0.1395364
#> [34,] 0.1527345 -0.1206612
#> [35,] 0.1696035 -0.1420207
#> [36,] 0.1722584 -0.1361803
#> [37,] 0.1746009 -0.1405353
#> [38,] 0.1643487 -0.1342771
#> [39,] 0.1486383 -0.1130450
#> [40,] 0.1721584 -0.1394931
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4233226 -0.4233226
#>  [2,]  0.4267824 -0.4267824
#>  [3,] -0.5754737  0.5754737
#>  [4,]  0.4286844 -0.4286844
#>  [5,] -0.5770879  0.5770879
#>  [6,]  0.4193466 -0.4193466
#>  [7,]  0.4231562 -0.4231562
#>  [8,] -0.5825029  0.5825029
#>  [9,]  0.4234962 -0.4234962
#> [10,]  0.4195130 -0.4195130
#> [11,]  0.4206354 -0.4206354
#> [12,] -0.5753192  0.5753192
#> [13,]  0.4251868 -0.4251868
#> [14,] -0.5828904  0.5828904
#> [15,] -0.5721403  0.5721403
#> [16,] -0.5752999  0.5752999
#> [17,]  0.4317633 -0.4317633
#> [18,] -0.5674341  0.5674341
#> [19,] -0.5771948  0.5771948
#> [20,]  0.4195914 -0.4195914
#> [21,]  0.4325040 -0.4325040
#> [22,]  0.4235941 -0.4235941
#> [23,]  0.4197473 -0.4197473
#> [24,] -0.5675297  0.5675297
#> [25,]  0.4231048 -0.4231048
#> [26,] -0.5825259  0.5825259
#> [27,]  0.4193503 -0.4193503
#> [28,]  0.4261006 -0.4261006
#> [29,] -0.5787737  0.5787737
#> [30,]  0.4230724 -0.4230724
#> [31,] -0.5773294  0.5773294
#> [32,]  0.4245006 -0.4245006
#> [33,] -0.5784350  0.5784350
#> [34,]  0.4320736 -0.4320736
#> [35,]  0.4227183 -0.4227183
#> [36,]  0.4234959 -0.4234959
#> [37,] -0.5781384  0.5781384
#> [38,] -0.5741065  0.5741065
#> [39,] -0.5650500  0.5650500
#> [40,] -0.5772883  0.5772883
#> 
#> $L
#> [1] 0.6908551 0.7162655 0.7231474 0.7361135 0.7237201 0.6895170
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536