Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12          V13
#>  [1,] -0.3542529 -0.02007271 -0.068848238 -0.012548079 -0.121137494
#>  [2,] -0.3167580 -0.04014807 -0.177279068 -0.153339915 -0.087056605
#>  [3,] -0.2760483 -0.09297679 -0.109830049 -0.114521618 -0.035593383
#>  [4,] -0.3701376 -0.10534680 -0.143966452 -0.181808082 -0.195758114
#>  [5,] -0.2255950 -0.15496626 -0.133466567 -0.141089653 -0.167112415
#>  [6,] -0.2407795 -0.19218957 -0.002502713 -0.003472625 -0.029382506
#>  [7,] -0.2942327 -0.07565017 -0.073311884 -0.079005882 -0.007848483
#>  [8,] -0.1629692 -0.13480807 -0.094179622 -0.019639728 -0.099973137
#>  [9,] -0.2838119 -0.20228999 -0.104894971 -0.141871327 -0.031886707
#> [10,] -0.3771255 -0.09049824 -0.051351354 -0.132961513 -0.012326913
#>                 V14          V15          V16           V17          V18
#>  [1,]  0.0006664751 -0.062462653 -0.063273772 -0.0133049808  0.056044397
#>  [2,] -0.0857709694  0.002470215  0.028634448 -0.0445629401  0.007989313
#>  [3,] -0.0851836487 -0.098446672 -0.094180925 -0.0009857225 -0.032733683
#>  [4,] -0.0308170099 -0.153884777 -0.121642390 -0.1024238338  0.053112718
#>  [5,] -0.1587154341 -0.103007397  0.008360338 -0.0381507375  0.062585399
#>  [6,] -0.1079436193 -0.041684267  0.062525481  0.0837759529 -0.006585063
#>  [7,]  0.0209785406 -0.032850495  0.025053901  0.0452491838 -0.027226436
#>  [8,] -0.0922252546 -0.038829791  0.051370497 -0.0724316690 -0.081709638
#>  [9,] -0.0221059949 -0.010601282 -0.116955070  0.0816555800  0.077464335
#> [10,] -0.1416262348 -0.155139589 -0.051785714 -0.0572049475  0.082039254
#>               V19         V2         V20          V21           V22
#>  [1,]  0.10752836 -0.2799983 0.040799681  0.163861588  0.1659878319
#>  [2,] -0.05907776 -0.3654909 0.068814054 -0.040004809 -0.0342040035
#>  [3,] -0.02362443 -0.3530490 0.043160899  0.075264358 -0.0641019138
#>  [4,]  0.09461435 -0.2610832 0.034029126 -0.001869767  0.1288254605
#>  [5,]  0.08780032 -0.2294615 0.057730778  0.029296012 -0.0009455423
#>  [6,]  0.03959130 -0.2416461 0.116089947  0.127271042  0.2372576662
#>  [7,]  0.14297873 -0.3016658 0.046401984  0.129938507  0.0599478700
#>  [8,]  0.02723783 -0.3169710 0.003043229  0.024565538 -0.0569256683
#>  [9,] -0.01317628 -0.2827224 0.135778612  0.114254802  0.1489657924
#> [10,]  0.05934543 -0.3585350 0.014238698 -0.027395653 -0.0108278628
#>                V23         V24         V25          V26        V27         V28
#>  [1,]  0.060876002  0.19972963  0.13545760  0.109932502 0.11124103  0.06466950
#>  [2,]  0.048569296 -0.02969307  0.03550329  0.010499053 0.11994434  0.07766478
#>  [3,]  0.107922910 -0.03553266 -0.08359256 -0.002175094 0.11008117  0.15620697
#>  [4,]  0.074419917  0.14301726  0.04705401  0.024594764 0.06084116  0.11003965
#>  [5,] -0.083649064 -0.02797506  0.07888657  0.104999467 0.02105954  0.06199489
#>  [6,]  0.254247550  0.22557465  0.16802766  0.195079051 0.09211926  0.04363065
#>  [7,]  0.129798967  0.06207400  0.21772871  0.263981805 0.21594879  0.26767842
#>  [8,]  0.003616278  0.03552954 -0.01108654 -0.011132555 0.05277023 -0.01348647
#>  [9,]  0.133446857  0.10079189  0.04393018  0.148018241 0.13648948  0.06382077
#> [10,] -0.013301899  0.12264852  0.10698459  0.167159229 0.05792063  0.13204486
#>               V29         V3           V30         V31         V32          V33
#>  [1,]  0.21472069 -0.2687246  0.0001187717  0.09851555 -0.03156224 -0.012193546
#>  [2,]  0.10248769 -0.2187348  0.1116073337  0.02876281  0.02133286  0.002868827
#>  [3,]  0.05103070 -0.3170149  0.0043538537  0.14498846  0.13868987  0.040485450
#>  [4,]  0.07095387 -0.3673118  0.1021983884  0.08391823 -0.09269189 -0.042468721
#>  [5,] -0.05043478 -0.2825320 -0.0349884204  0.01767501 -0.12873985 -0.095098630
#>  [6,]  0.13067528 -0.2277013  0.0866093413  0.02814915  0.02135370  0.048997598
#>  [7,]  0.12164610 -0.3320055  0.2080508567  0.05487910  0.07797044  0.040564391
#>  [8,]  0.06424234 -0.1973107 -0.0281250189 -0.02700301 -0.15963348 -0.006873609
#>  [9,]  0.12400570 -0.3125352  0.0765930820 -0.12919769  0.02593041 -0.084276007
#> [10,]  0.04877361 -0.2743675  0.1045345314  0.08379930 -0.07933675  0.102171995
#>               V34          V35          V36          V37          V38
#>  [1,] -0.12647952 -0.124605590 -0.189740092 -0.123169338 -0.138126356
#>  [2,] -0.03064448 -0.056869386 -0.189617060 -0.167344936 -0.100545885
#>  [3,]  0.09391307  0.025866488  0.057097739 -0.067472256 -0.045863648
#>  [4,] -0.03129512 -0.100731355 -0.103691387 -0.082630253 -0.127914885
#>  [5,] -0.02676046  0.003754972  0.009725709 -0.086228386 -0.144105042
#>  [6,] -0.04296907  0.078504557  0.044932996 -0.023946558  0.050004420
#>  [7,]  0.06639748 -0.004235853  0.056071034  0.051743876  0.009078498
#>  [8,]  0.04898082  0.019175429 -0.013884487 -0.066932828 -0.034435526
#>  [9,]  0.02016496 -0.159657511 -0.069141976  0.003521771 -0.148646914
#> [10,] -0.05668971 -0.002251892  0.046899062  0.064754798 -0.067794328
#>                V39         V4          V40         V41         V42         V43
#>  [1,] -0.146618866 -0.3085235  0.007645316 -0.08784751 -0.03307666 -0.01631427
#>  [2,] -0.001738119 -0.1545615 -0.192460905 -0.19744747 -0.11419046 -0.15635201
#>  [3,]  0.020029136 -0.3040688 -0.069522260 -0.04261366 -0.06690406 -0.06990377
#>  [4,]  0.022948397 -0.2456597 -0.084994041 -0.08726255 -0.09633162 -0.13195929
#>  [5,] -0.142383795 -0.2504204 -0.145873323 -0.10962497 -0.11093205 -0.15256880
#>  [6,]  0.007926089 -0.2265078  0.023517205 -0.14660071 -0.01590582 -0.12082322
#>  [7,] -0.074879242 -0.3613756  0.096034970  0.02096054 -0.03861415 -0.06829813
#>  [8,] -0.106274068 -0.2662741 -0.069450749 -0.10753165 -0.17799608 -0.17080341
#>  [9,] -0.121381559 -0.2504037 -0.236111499 -0.15347715 -0.08453530 -0.08365356
#> [10,] -0.006104907 -0.2746257 -0.025369951  0.06723514  0.02897983 -0.03276883
#>                V44         V45          V46         V47         V48        V49
#>  [1,] -0.036181854 -0.10815128 -0.168585189 -0.13516596 -0.17469597 -0.2061143
#>  [2,] -0.085864255 -0.05041777 -0.150493969 -0.17502328 -0.28619620 -0.2628314
#>  [3,]  0.022343320 -0.07702716 -0.006831014 -0.11369584 -0.25356925 -0.3214500
#>  [4,] -0.092848626 -0.07910578 -0.116307413 -0.13842483 -0.24514018 -0.2330833
#>  [5,] -0.158450945 -0.15659125 -0.200979375 -0.14707623 -0.09494709 -0.1839398
#>  [6,] -0.075675168 -0.10428842 -0.169659464 -0.17904662 -0.30527175 -0.3034574
#>  [7,] -0.189484486 -0.07860405 -0.164325143 -0.30487103 -0.14902311 -0.2003182
#>  [8,] -0.102093033 -0.19992898 -0.152384163 -0.22327654 -0.17529621 -0.3132301
#>  [9,] -0.284120461 -0.25978331 -0.202008138 -0.22437553 -0.16683283 -0.2447631
#> [10,] -0.009880636 -0.07437271 -0.156353155 -0.08842694 -0.25929983 -0.2698353
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.1826567 -0.2248449 -0.2290360 -0.3811299 -0.3739977 -0.4035733
#>  [2,] -0.2347980 -0.2754592 -0.2855160 -0.2451571 -0.3788151 -0.3879963
#>  [3,] -0.3181561 -0.2185678 -0.3868186 -0.2854473 -0.2320363 -0.2615614
#>  [4,] -0.3283204 -0.4084530 -0.2899557 -0.3738907 -0.4010506 -0.3888385
#>  [5,] -0.2925589 -0.2390711 -0.3025285 -0.2444929 -0.2193128 -0.2235981
#>  [6,] -0.2477979 -0.3334588 -0.3752745 -0.4645480 -0.3700538 -0.3613819
#>  [7,] -0.2200329 -0.3343520 -0.3182166 -0.3745701 -0.3936560 -0.4633774
#>  [8,] -0.1381692 -0.1870622 -0.2562765 -0.2337834 -0.3005666 -0.1717632
#>  [9,] -0.2413138 -0.3921390 -0.4240030 -0.3202140 -0.3106575 -0.3191413
#> [10,] -0.1780536 -0.3254828 -0.2680833 -0.3329805 -0.2884080 -0.3853252
#>              V55        V56        V57        V58        V59          V6
#>  [1,] -0.4046604 -0.3514941 -0.3085674 -0.2786299 -0.3004603 -0.17195554
#>  [2,] -0.2608932 -0.2825365 -0.3030259 -0.2720019 -0.2791011 -0.22891842
#>  [3,] -0.3707168 -0.3388044 -0.2330372 -0.2983348 -0.2955108 -0.13860210
#>  [4,] -0.2842556 -0.3017845 -0.3070652 -0.3923547 -0.3501381 -0.26715447
#>  [5,] -0.3446678 -0.2882864 -0.3897559 -0.3309800 -0.3561761 -0.26490711
#>  [6,] -0.3211670 -0.3415988 -0.4070897 -0.3203219 -0.3130305 -0.11993060
#>  [7,] -0.3936540 -0.3888618 -0.3965206 -0.3772844 -0.4564659 -0.25709852
#>  [8,] -0.2698991 -0.1890562 -0.3008618 -0.2066452 -0.1692026 -0.08441328
#>  [9,] -0.3536830 -0.3322414 -0.3689783 -0.3644152 -0.2935415 -0.13353359
#> [10,] -0.3526877 -0.3915114 -0.3869465 -0.4069927 -0.3992650 -0.25125081
#>              V60         V7         V8          V9
#>  [1,] -0.2989442 -0.1869767 -0.1139834 -0.11161516
#>  [2,] -0.3689587 -0.1361195 -0.1797394 -0.08514477
#>  [3,] -0.3546934 -0.1046122 -0.1431772 -0.10080580
#>  [4,] -0.3927403 -0.1906459 -0.2306876 -0.21960994
#>  [5,] -0.3913035 -0.1484825 -0.2494795 -0.19384928
#>  [6,] -0.4017393 -0.2647276 -0.1863157 -0.11848373
#>  [7,] -0.2806973 -0.1226605 -0.1710961 -0.06624574
#>  [8,] -0.3346649 -0.1648576 -0.1171548 -0.06759838
#>  [9,] -0.4467931 -0.3166430 -0.1962968 -0.26868092
#> [10,] -0.3377999 -0.2101464 -0.1546133 -0.11134655
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]         [,6]
#> [1,]  0.01316231 -0.04993576 -0.083146806 -0.09390333 -0.01429196 -0.031726698
#> [2,] -0.10966342  0.11998416  0.005480188  0.04023868  0.01533011 -0.001735411
#>             [,7]        [,8]         [,9]       [,10]
#> [1,]  0.04791387 -0.08491771 -0.112456796 -0.05534100
#> [2,] -0.02139518  0.04945357 -0.007888564 -0.04773115
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12         V13
#>  [1,] 7.058272e-05 0.0005432935 0.0004668075 0.0007687692 0.001321461
#>  [2,] 2.559124e-04 0.0020056566 0.0022833329 0.0023675477 0.002285462
#>  [3,] 2.168910e-04 0.0016163221 0.0017338512 0.0018975801 0.002123793
#>  [4,] 2.562401e-04 0.0019436456 0.0021679865 0.0022830043 0.002332606
#>  [5,] 1.084125e-04 0.0007649806 0.0007316369 0.0008956032 0.001212426
#>  [6,] 1.514013e-04 0.0010910582 0.0011367538 0.0013114345 0.001559485
#>  [7,] 6.066494e-05 0.0005381958 0.0005136129 0.0007228878 0.001034487
#>  [8,] 1.583588e-04 0.0012468611 0.0013955743 0.0014728247 0.001439217
#>  [9,] 2.313640e-04 0.0017246766 0.0018928154 0.0020299835 0.002173949
#> [10,] 1.436764e-04 0.0010681148 0.0010650413 0.0012836080 0.001688270
#>               V14         V15         V16         V17         V18         V19
#>  [1,] 0.002246838 0.002830569 0.003390087 0.003627186 0.003769881 0.003168072
#>  [2,] 0.001862173 0.001578015 0.001961009 0.002258019 0.002720436 0.003952181
#>  [3,] 0.002302156 0.002436810 0.002963751 0.003219661 0.003544533 0.004086565
#>  [4,] 0.002140854 0.002008795 0.002465113 0.002758217 0.003182050 0.004185828
#>  [5,] 0.001694918 0.002012634 0.002425277 0.002564620 0.002706400 0.002667736
#>  [6,] 0.001835393 0.002032453 0.002450935 0.002610054 0.002810609 0.003003979
#>  [7,] 0.001520932 0.001839338 0.002206547 0.002333234 0.002449121 0.002217441
#>  [8,] 0.001237060 0.001116018 0.001390622 0.001574120 0.001872197 0.002609031
#>  [9,] 0.002181373 0.002176896 0.002652104 0.002916689 0.003276165 0.004014118
#> [10,] 0.002274542 0.002686353 0.003239760 0.003443545 0.003621687 0.003566914
#>                 V2         V20         V21         V22         V23         V24
#>  [1,] 0.0001189277 0.002728565 0.002881076 0.003200474 0.004060757 0.004829648
#>  [2,] 0.0003864021 0.004942885 0.005269609 0.005096620 0.004932001 0.004856652
#>  [3,] 0.0003309455 0.004636722 0.004934408 0.004884235 0.005145282 0.005512056
#>  [4,] 0.0003833411 0.005045937 0.005370676 0.005248231 0.005249438 0.005339798
#>  [5,] 0.0001629658 0.002795894 0.002970874 0.003049689 0.003538811 0.004056569
#>  [6,] 0.0002300471 0.003221265 0.003407857 0.003392237 0.003709798 0.004094449
#>  [7,] 0.0001083116 0.002031001 0.002112628 0.002242986 0.002734847 0.003191543
#>  [8,] 0.0002453534 0.003222820 0.003434426 0.003303726 0.003266478 0.003301110
#>  [9,] 0.0003448916 0.004685471 0.004977463 0.004915731 0.005059729 0.005274015
#> [10,] 0.0002234354 0.003653673 0.003858913 0.003943375 0.004484478 0.005063824
#>               V25         V26         V27         V28         V29           V3
#>  [1,] 0.005349651 0.005519714 0.005030555 0.005036160 0.005276042 0.0002317256
#>  [2,] 0.004442073 0.004486883 0.004838555 0.005008249 0.004514873 0.0003523864
#>  [3,] 0.005485712 0.005605569 0.005543668 0.005609842 0.005427672 0.0003530470
#>  [4,] 0.005076264 0.005168393 0.005373321 0.005513670 0.005126268 0.0003727278
#>  [5,] 0.004297151 0.004430003 0.004155628 0.004153859 0.004220930 0.0002119045
#>  [6,] 0.004193405 0.004333853 0.004219402 0.004356111 0.004363124 0.0002659995
#>  [7,] 0.003443005 0.003556962 0.003294875 0.003332389 0.003432443 0.0001733312
#>  [8,] 0.003060806 0.003086458 0.003255927 0.003372344 0.003105133 0.0002269683
#>  [9,] 0.005154011 0.005265553 0.005343455 0.005453161 0.005169743 0.0003551413
#> [10,] 0.005309503 0.005449516 0.005144933 0.005160583 0.005191530 0.0002866896
#>               V30         V31         V32         V33         V34         V35
#>  [1,] 0.005084547 0.004856174 0.003558883 0.004069804 0.004845414 0.005247120
#>  [2,] 0.003808078 0.002990653 0.003180524 0.002706342 0.002299370 0.002151949
#>  [3,] 0.004915928 0.004283915 0.003828945 0.003734396 0.003837479 0.003921405
#>  [4,] 0.004479856 0.003705140 0.003639379 0.003301408 0.003086012 0.003010256
#>  [5,] 0.003996108 0.003698859 0.003002348 0.003196742 0.003590392 0.003748292
#>  [6,] 0.004035149 0.003601436 0.003057220 0.003118901 0.003351660 0.003480680
#>  [7,] 0.003242863 0.003068545 0.002302420 0.002533341 0.002952856 0.003166086
#>  [8,] 0.002662606 0.002153474 0.002200928 0.001960451 0.001789856 0.001755037
#>  [9,] 0.004611527 0.003933918 0.003668868 0.003470491 0.003428608 0.003420339
#> [10,] 0.004877283 0.004490395 0.003616694 0.003799064 0.004256482 0.004492541
#>               V36         V37         V38         V39           V4         V40
#>  [1,] 0.005269131 0.004180979 0.003099969 0.002711795 0.0001974515 0.002846968
#>  [2,] 0.001909402 0.001974522 0.002377814 0.002762169 0.0003948765 0.002757881
#>  [3,] 0.003773889 0.003311149 0.003059485 0.003109333 0.0003584756 0.003179517
#>  [4,] 0.002782144 0.002628024 0.002786164 0.003049278 0.0004022046 0.003062878
#>  [5,] 0.003641471 0.002952960 0.002345324 0.002139467 0.0002004116 0.002242963
#>  [6,] 0.003414635 0.002925763 0.002588744 0.002556165 0.0002583736 0.002590749
#>  [7,] 0.003217767 0.002617070 0.002057439 0.001911350 0.0001607694 0.002026617
#>  [8,] 0.001601434 0.001545149 0.001685557 0.001873787 0.0002488656 0.001898747
#>  [9,] 0.003211187 0.002888563 0.002833327 0.002979763 0.0003722504 0.003017220
#> [10,] 0.004428397 0.003657746 0.002981794 0.002788466 0.0002735449 0.002919900
#>               V41         V42         V43          V44           V45
#>  [1,] 0.002298730 0.001745248 0.001330271 0.0007325736 -2.184533e-05
#>  [2,] 0.002416526 0.002518448 0.002339625 0.0023974552  2.450303e-03
#>  [3,] 0.002662994 0.002460814 0.002166904 0.0019832356  1.709186e-03
#>  [4,] 0.002649726 0.002646239 0.002397609 0.0023463512  2.287619e-03
#>  [5,] 0.001810675 0.001494996 0.001267514 0.0010225953  6.061611e-04
#>  [6,] 0.002186340 0.001945743 0.001642637 0.0013773312  1.084640e-03
#>  [7,] 0.001676469 0.001332986 0.001054891 0.0007278668  2.680970e-04
#>  [8,] 0.001608688 0.001601607 0.001491357 0.0015196432  1.464321e-03
#>  [9,] 0.002586085 0.002498548 0.002227880 0.0020911670  1.931733e-03
#> [10,] 0.002388049 0.001991202 0.001667612 0.0013318296  8.303299e-04
#>                V46          V47          V48          V49           V5
#>  [1,] 5.850594e-05 0.0001771203 6.257854e-05 2.299229e-05 0.0003754463
#>  [2,] 1.855651e-03 0.0013187093 1.061095e-03 6.151075e-04 0.0005829116
#>  [3,] 1.306174e-03 0.0009737387 7.566715e-04 4.274229e-04 0.0005597973
#>  [4,] 1.742294e-03 0.0012533131 9.950983e-04 5.724244e-04 0.0006036481
#>  [5,] 4.450925e-04 0.0003608803 2.631558e-04 1.376398e-04 0.0003310733
#>  [6,] 8.728784e-04 0.0006739307 4.988595e-04 2.825370e-04 0.0004132002
#>  [7,] 2.338143e-04 0.0002481715 1.540929e-04 8.508656e-05 0.0002921082
#>  [8,] 1.093897e-03 0.0007851589 6.340478e-04 3.652484e-04 0.0003706512
#>  [9,] 1.482031e-03 0.0010792541 8.439759e-04 4.837435e-04 0.0005652067
#> [10,] 6.513651e-04 0.0005442136 3.904964e-04 2.131450e-04 0.0004586303
#>                V50          V51          V52          V53          V54
#>  [1,] 1.043962e-04 4.008084e-05 1.805275e-05 5.921570e-05 7.373001e-05
#>  [2,] 1.640902e-04 1.693131e-04 1.324988e-04 8.095201e-05 8.737661e-05
#>  [3,] 1.595426e-04 1.332389e-04 9.780418e-05 8.142786e-05 9.371361e-05
#>  [4,] 1.731563e-04 1.633764e-04 1.260656e-04 8.587124e-05 9.500587e-05
#>  [5,] 9.301836e-05 6.031732e-05 3.979256e-05 5.318980e-05 6.421871e-05
#>  [6,] 1.250207e-04 9.261077e-05 6.366358e-05 6.202641e-05 7.206643e-05
#>  [7,] 7.706277e-05 4.063039e-05 2.131100e-05 4.234945e-05 4.938122e-05
#>  [8,] 1.011911e-04 1.035800e-04 7.860850e-05 5.219620e-05 5.842354e-05
#>  [9,] 1.635731e-04 1.436307e-04 1.088472e-04 8.156394e-05 9.231101e-05
#> [10,] 1.272504e-04 8.237025e-05 5.468011e-05 6.756485e-05 8.097988e-05
#>                V55          V56          V57          V58          V59
#>  [1,] 7.538915e-05 5.451232e-05 6.894982e-05 3.796375e-05 6.465659e-05
#>  [2,] 6.569544e-05 7.345071e-05 4.729391e-05 6.644875e-05 5.804657e-05
#>  [3,] 8.038635e-05 7.716105e-05 6.332317e-05 6.053999e-05 6.758689e-05
#>  [4,] 7.639697e-05 7.896531e-05 5.674498e-05 6.821501e-05 6.571981e-05
#>  [5,] 5.899267e-05 5.054646e-05 5.138093e-05 3.319118e-05 4.819202e-05
#>  [6,] 6.588366e-05 5.866538e-05 5.286361e-05 4.709689e-05 5.649212e-05
#>  [7,] 5.091839e-05 3.965460e-05 4.380707e-05 2.773567e-05 4.219203e-05
#>  [8,] 4.316847e-05 4.875826e-05 3.346199e-05 4.176446e-05 3.906897e-05
#>  [9,] 7.711769e-05 7.594795e-05 5.957025e-05 6.338968e-05 6.567321e-05
#> [10,] 7.589910e-05 6.528017e-05 6.462437e-05 4.544648e-05 6.262918e-05
#>                 V6          V60           V7           V8           V9
#>  [1,] 0.0006455261 6.390427e-05 0.0009778488 0.0009096395 0.0004686031
#>  [2,] 0.0009859085 4.018982e-05 0.0009490336 0.0009718609 0.0016057509
#>  [3,] 0.0009898372 5.591296e-05 0.0011094303 0.0010960264 0.0013340331
#>  [4,] 0.0010262357 4.866569e-05 0.0010542488 0.0010669065 0.0015727352
#>  [5,] 0.0006279069 4.443049e-05 0.0007931698 0.0007365597 0.0006639940
#>  [6,] 0.0007164459 4.724487e-05 0.0008697740 0.0008688390 0.0009032657
#>  [7,] 0.0004933147 4.146084e-05 0.0006877383 0.0006657022 0.0004524693
#>  [8,] 0.0006626074 2.859639e-05 0.0006549281 0.0006503728 0.0010032858
#>  [9,] 0.0009641841 5.098904e-05 0.0010435363 0.0010470085 0.0014074932
#> [10,] 0.0008240833 5.784626e-05 0.0010473317 0.0010040636 0.0009094905
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]       [,5]        [,6]
#> [1,]  0.01688567  0.02068479  0.02221899  0.01991412  0.0263927  0.01970074
#> [2,] -0.01688567 -0.02068479 -0.02221899 -0.01991412 -0.0263927 -0.01970074
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.02122143  0.02174295  0.02302321  0.02414047
#> [2,] -0.02122143 -0.02174295 -0.02302321 -0.02414047
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.69613255 0.38689257 0.59940801 0.50090565 0.34082478 0.64678006
#>  [7] 0.78898388 0.04893366 0.47461851 0.59477234
#> 
#> $B[[2]]
#> [1]  0.02916695 -0.04262155
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.008553075 0.006758997 0.008517367 0.007828543 0.006671876 0.006889049
#>  [7] 0.005666827 0.004720534 0.007989142 0.008373429
#> 
#> $vB[[2]]
#> [1]  0.01921747 -0.01921747
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#>  [2,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [3,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#>  [4,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#>  [5,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#>  [6,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#>  [7,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#>  [8,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#>  [9,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [10,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [11,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [12,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [13,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [14,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [15,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [16,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [17,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [18,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [19,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [20,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [21,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [22,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [23,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [24,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [25,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [26,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [27,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [28,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [29,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [30,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [31,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [32,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [33,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [34,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [35,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [36,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [37,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [38,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [39,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [40,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#>  [2,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [3,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#>  [4,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#>  [5,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#>  [6,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#>  [7,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#>  [8,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#>  [9,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [10,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [11,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [12,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [13,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [14,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [15,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [16,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [17,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [18,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [19,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [20,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [21,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [22,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [23,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [24,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [25,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [26,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [27,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [28,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [29,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [30,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [31,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [32,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [33,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [34,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [35,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [36,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [37,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [38,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [39,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [40,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#>  [2,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [3,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#>  [4,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#>  [5,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#>  [6,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#>  [7,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#>  [8,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#>  [9,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [10,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [11,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [12,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [13,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [14,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [15,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [16,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [17,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [18,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [19,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [20,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [21,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [22,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [23,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [24,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [25,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [26,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [27,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [28,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [29,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [30,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [31,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [32,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [33,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [34,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [35,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [36,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [37,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [38,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [39,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [40,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#>  [2,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [3,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#>  [4,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#>  [5,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#>  [6,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#>  [7,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#>  [8,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#>  [9,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [10,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [11,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [12,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [13,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [14,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [15,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [16,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [17,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [18,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [19,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [20,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [21,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [22,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [23,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [24,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [25,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [26,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [27,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [28,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [29,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [30,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [31,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [32,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [33,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [34,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [35,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [36,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [37,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [38,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [39,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [40,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#>  [2,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [3,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#>  [4,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#>  [5,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#>  [6,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#>  [7,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#>  [8,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#>  [9,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [10,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [11,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [12,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [13,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [14,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [15,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [16,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [17,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [18,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [19,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [20,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [21,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [22,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [23,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [24,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [25,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [26,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [27,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [28,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [29,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [30,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [31,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [32,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [33,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [34,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [35,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [36,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [37,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [38,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [39,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [40,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#>  [2,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [3,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#>  [4,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#>  [5,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#>  [6,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#>  [7,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#>  [8,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#>  [9,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [10,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [11,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [12,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [13,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [14,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [15,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [16,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [17,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [18,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [19,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [20,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [21,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [22,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [23,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [24,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [25,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [26,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [27,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [28,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [29,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [30,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [31,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [32,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [33,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [34,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [35,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [36,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [37,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [38,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [39,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [40,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7619559 0.5649006 0.6506409 0.6300876 0.5182090 0.8026036 0.8578401
#>  [2,] 0.7431627 0.5258611 0.7181445 0.6237062 0.4568954 0.8585955 0.9005356
#>  [3,] 0.7889164 0.5296934 0.6788626 0.6288598 0.4419642 0.8727509 0.9027059
#>  [4,] 0.7962395 0.5779047 0.6953080 0.6548632 0.4748279 0.8522756 0.9006851
#>  [5,] 0.6629559 0.5148847 0.6715698 0.5863924 0.4339793 0.7355080 0.8115757
#>  [6,] 0.7592718 0.4615894 0.6331299 0.5360538 0.3301742 0.8269399 0.8748457
#>  [7,] 0.5180421 0.3465934 0.5590722 0.3637375 0.2892298 0.6799173 0.7473430
#>  [8,] 0.7981792 0.5938605 0.6992825 0.6546208 0.4722935 0.8433648 0.8906088
#>  [9,] 0.7707961 0.5526614 0.6705282 0.6174456 0.4492783 0.8092942 0.8743366
#> [10,] 0.7379196 0.5103217 0.6586826 0.5806447 0.4176331 0.8048896 0.8657269
#> [11,] 0.7009758 0.4083698 0.5223647 0.4491992 0.3444546 0.7715274 0.7942319
#> [12,] 0.7239312 0.4486107 0.6337009 0.5034668 0.3175508 0.7964896 0.8626595
#> [13,] 0.6089335 0.4155552 0.6629488 0.4635500 0.3134742 0.7476013 0.8560630
#> [14,] 0.6798113 0.4653143 0.7082346 0.5535124 0.3889794 0.8514816 0.8993303
#> [15,] 0.8056576 0.5244626 0.6250642 0.6079200 0.4464181 0.8594544 0.8896534
#> [16,] 0.7765488 0.5553055 0.6402788 0.5930598 0.4450688 0.8449908 0.8702465
#> [17,] 0.8153741 0.5439922 0.6523891 0.6371306 0.4604581 0.8551879 0.8873362
#> [18,] 0.7778046 0.5217563 0.6192188 0.5733742 0.4358064 0.8401554 0.8763565
#> [19,] 0.7627937 0.5606383 0.7323824 0.6540840 0.4750490 0.8476751 0.8928442
#> [20,] 0.6020858 0.3862043 0.6529343 0.4586806 0.2794425 0.7227177 0.8368608
#> [21,] 0.6809020 0.3896005 0.5680101 0.4839756 0.3547549 0.7996713 0.8459644
#> [22,] 0.8264113 0.5811349 0.6814878 0.6711603 0.4964103 0.8676206 0.8998754
#> [23,] 0.7546457 0.5769225 0.7101937 0.6177559 0.4375913 0.8268829 0.8854835
#> [24,] 0.6822152 0.3974122 0.5492414 0.4978529 0.3355493 0.7644061 0.7970803
#> [25,] 0.7487283 0.5762473 0.6612965 0.6159786 0.5256269 0.7878897 0.8576247
#> [26,] 0.7732208 0.5595596 0.6694349 0.6201154 0.4527232 0.8074596 0.8703477
#> [27,] 0.4920908 0.3267850 0.5657402 0.3731459 0.2838369 0.6760321 0.7355480
#> [28,] 0.6763152 0.4486671 0.6927629 0.5317968 0.3891477 0.8578543 0.9103097
#> [29,] 0.7512936 0.4774806 0.6513398 0.5862818 0.4596088 0.8461960 0.8602558
#> [30,] 0.7617516 0.4655829 0.6287542 0.5485174 0.3376867 0.8187067 0.8617802
#> [31,] 0.6518768 0.3882368 0.6469026 0.4520418 0.2660577 0.7504088 0.8205501
#> [32,] 0.7364078 0.4769615 0.6840560 0.6007646 0.4373558 0.8590478 0.8989857
#> [33,] 0.6549132 0.4322471 0.6680196 0.4885955 0.3145568 0.7765326 0.8657580
#> [34,] 0.8039969 0.5829562 0.6905979 0.6616475 0.4908708 0.8524161 0.9001440
#> [35,] 0.8277378 0.5807600 0.6681055 0.6544340 0.5235503 0.8892577 0.9069843
#> [36,] 0.7212289 0.4435496 0.6225820 0.5561397 0.4238143 0.8279602 0.8439759
#> [37,] 0.7590962 0.5332885 0.6845897 0.6116741 0.4314774 0.8309761 0.8628654
#> [38,] 0.7603839 0.5364649 0.6514010 0.6065476 0.4739201 0.8277624 0.8403963
#> [39,] 0.8064407 0.5054507 0.6211103 0.6128757 0.4316369 0.8656302 0.8901814
#> [40,] 0.7149412 0.4283094 0.6114202 0.5327213 0.4012801 0.8391831 0.8458694
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4386346 0.6541100 0.7277980
#>  [2,] 0.4093326 0.6452869 0.7533912
#>  [3,] 0.3984807 0.6789362 0.7347936
#>  [4,] 0.3977671 0.6846706 0.7459582
#>  [5,] 0.4043292 0.5463424 0.6918835
#>  [6,] 0.3026826 0.5633482 0.6685225
#>  [7,] 0.3005845 0.3749701 0.5452764
#>  [8,] 0.4057718 0.6777355 0.7383846
#>  [9,] 0.3890709 0.6430276 0.7174290
#> [10,] 0.3698605 0.5896901 0.6943482
#> [11,] 0.3258564 0.5448936 0.5648343
#> [12,] 0.2885565 0.5355898 0.6648094
#> [13,] 0.2928916 0.4160148 0.6979599
#> [14,] 0.3471531 0.5722010 0.7439586
#> [15,] 0.3720981 0.6882684 0.7019848
#> [16,] 0.3994201 0.6702020 0.6868600
#> [17,] 0.3899910 0.6831628 0.7123630
#> [18,] 0.3621785 0.6663644 0.6696887
#> [19,] 0.4176117 0.6609211 0.7500410
#> [20,] 0.2580954 0.3786669 0.6856329
#> [21,] 0.3055304 0.5250766 0.6508977
#> [22,] 0.4180927 0.7175851 0.7320159
#> [23,] 0.4087074 0.6394472 0.7312993
#> [24,] 0.3037987 0.5367127 0.6249604
#> [25,] 0.4433837 0.6502916 0.7230847
#> [26,] 0.3930035 0.6486622 0.7152572
#> [27,] 0.2850808 0.3402951 0.5443130
#> [28,] 0.3553629 0.5741112 0.7405847
#> [29,] 0.3808712 0.6216449 0.7069937
#> [30,] 0.3217830 0.5675540 0.6622039
#> [31,] 0.2503502 0.4288806 0.6336796
#> [32,] 0.3804229 0.6293349 0.7514798
#> [33,] 0.2955753 0.4522573 0.7059783
#> [34,] 0.4087963 0.6953903 0.7473687
#> [35,] 0.4201814 0.7476800 0.7189537
#> [36,] 0.3576014 0.6067539 0.6789447
#> [37,] 0.3753940 0.6391642 0.6864496
#> [38,] 0.3959863 0.6629795 0.6616577
#> [39,] 0.3586851 0.6743110 0.7065600
#> [40,] 0.3457051 0.5938100 0.6627161
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5338543 0.4661457
#>  [2,] 0.5357931 0.4642069
#>  [3,] 0.5373379 0.4626621
#>  [4,] 0.5354221 0.4645779
#>  [5,] 0.5324441 0.4675559
#>  [6,] 0.5426854 0.4573146
#>  [7,] 0.5392469 0.4607531
#>  [8,] 0.5344772 0.4655228
#>  [9,] 0.5359223 0.4640777
#> [10,] 0.5375338 0.4624662
#> [11,] 0.5417809 0.4582191
#> [12,] 0.5426852 0.4573148
#> [13,] 0.5415368 0.4584632
#> [14,] 0.5389480 0.4610520
#> [15,] 0.5389989 0.4610011
#> [16,] 0.5359732 0.4640268
#> [17,] 0.5375498 0.4624502
#> [18,] 0.5386160 0.4613840
#> [19,] 0.5339929 0.4660071
#> [20,] 0.5428476 0.4571524
#> [21,] 0.5434597 0.4565403
#> [22,] 0.5353218 0.4646782
#> [23,] 0.5341062 0.4658938
#> [24,] 0.5412020 0.4587980
#> [25,] 0.5330926 0.4669074
#> [26,] 0.5354574 0.4645426
#> [27,] 0.5390531 0.4609469
#> [28,] 0.5400958 0.4599042
#> [29,] 0.5386527 0.4613473
#> [30,] 0.5415163 0.4584837
#> [31,] 0.5433957 0.4566043
#> [32,] 0.5387168 0.4612832
#> [33,] 0.5419862 0.4580138
#> [34,] 0.5351420 0.4648580
#> [35,] 0.5357121 0.4642879
#> [36,] 0.5393467 0.4606533
#> [37,] 0.5357482 0.4642518
#> [38,] 0.5347313 0.4652687
#> [39,] 0.5400599 0.4599401
#> [40,] 0.5405019 0.4594981
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]       [,3]         [,4]        [,5]      [,6]
#>  [1,]  1.16343281  0.26107540 0.62185734  0.532592674  0.07286830 1.4026472
#>  [2,]  1.06247234  0.10353685 0.93527620  0.505309348 -0.17284745 1.8036732
#>  [3,]  1.31840600  0.11891340 0.74854972  0.527328400 -0.23319415 1.9255032
#>  [4,]  1.36295454  0.31417788 0.82505329  0.640485534 -0.10077341 1.7525618
#>  [5,]  0.67649447  0.05955645 0.71529374  0.349071390 -0.26563400 1.0227503
#>  [6,]  1.14869115 -0.15394578 0.54566777  0.144465948 -0.70739741 1.5640931
#>  [7,]  0.07219977 -0.63404712 0.23739736 -0.559178853 -0.89912747 0.7533919
#>  [8,]  1.37495286  0.37994766 0.84388362  0.639413347 -0.11093949 1.6834798
#>  [9,]  1.21281207  0.21142965 0.71057488  0.478720187 -0.20358696 1.4454306
#> [10,]  1.03518349  0.04129248 0.65742881  0.325420708 -0.33249756 1.4171398
#> [11,]  0.85194908 -0.37070882 0.08951855 -0.203906875 -0.64350528 1.2169560
#> [12,]  0.96404636 -0.20628572 0.54812677  0.013867255 -0.76505050 1.3644971
#> [13,]  0.44283154 -0.34104671 0.67646281 -0.146059152 -0.78392696 1.0858601
#> [14,]  0.75290490 -0.13896604 0.88682557  0.214872674 -0.45160449 1.7462687
#> [15,]  1.42203741  0.09792877 0.51109962  0.438577329 -0.21515359 1.8107661
#> [16,]  1.24566671  0.22213079 0.57657430  0.376629060 -0.22061536 1.6958410
#> [17,]  1.48531525  0.17642502 0.62955749  0.562931795 -0.15849866 1.7758839
#> [18,]  1.25291822  0.08708025 0.48623370  0.295631093 -0.25819941 1.6593849
#> [19,]  1.16805738  0.24375283 1.00674385  0.637039909 -0.09988700 1.7164819
#> [20,]  0.41416356 -0.46329556 0.63196231 -0.165655552 -0.94722867 0.9579827
#> [21,]  0.75792030 -0.44899198 0.27373683 -0.064119521 -0.59820310 1.3842413
#> [22,]  1.56040379  0.32743380 0.76061778  0.713437467 -0.01435917 1.8800821
#> [23,]  1.12354491  0.31015273 0.89632485  0.480033855 -0.25094324 1.5636946
#> [24,]  0.76397086 -0.41625919 0.19760614 -0.008588343 -0.68319163 1.1769896
#> [25,]  1.09184159  0.30738697 0.66907727  0.472513937  0.10259755 1.3122519
#> [26,]  1.22658803  0.23937496 0.70563021  0.490038065 -0.18967383 1.4335867
#> [27,] -0.03163938 -0.72276227 0.26449192 -0.518744398 -0.92550836 0.7355962
#> [28,]  0.73688924 -0.20605764 0.81306826  0.127358964 -0.45089639 1.7975817
#> [29,]  1.10552364 -0.09013876 0.62493359  0.348615541 -0.16191745 1.7050719
#> [30,]  1.16230657 -0.13788646 0.52687596  0.194682246 -0.67362019 1.5076097
#> [31,]  0.62729934 -0.45472981 0.60545185 -0.192424547 -1.01471726 1.1007940
#> [32,]  1.02738097 -0.09221934 0.77247470  0.408651801 -0.25190026 1.8074038
#> [33,]  0.64070666 -0.27268880 0.69924177 -0.045625757 -0.77890117 1.2455729
#> [34,]  1.41146489  0.33492083 0.80291582  0.670644817 -0.03652097 1.7536781
#> [35,]  1.56967857  0.32589399 0.69962892  0.638587063  0.09427110 2.0831816
#> [36,]  0.95056553 -0.22676840 0.50052186  0.225509794 -0.30713448 1.5712391
#> [37,]  1.14773064  0.13335129 0.77494555  0.454354625 -0.27582597 1.5925612
#> [38,]  1.15478516  0.14611917 0.62520319  0.432823003 -0.10441418 1.5698514
#> [39,]  1.42704649  0.02180351 0.49426349  0.459416330 -0.27517555 1.8628625
#> [40,]  0.91950473 -0.28875231 0.45328578  0.131072338 -0.40013410 1.6521627
#>           [,7]       [,8]       [,9]     [,10]
#>  [1,] 1.797465 -0.2467053  0.6371549 0.9834791
#>  [2,] 2.203190 -0.3667251  0.5983856 1.1167816
#>  [3,] 2.227659 -0.4117996  0.7488875 1.0190813
#>  [4,] 2.204860 -0.4147777  0.7753203 1.0771706
#>  [5,] 1.460281 -0.3874587  0.1859033 0.8089395
#>  [6,] 1.944500 -0.8345559  0.2547617 0.7015099
#>  [7,] 1.084492 -0.8445162 -0.5109531 0.1816030
#>  [8,] 2.096975 -0.3814726  0.7433844 1.0375895
#>  [9,] 1.939858 -0.4512193  0.5885291 0.9317440
#> [10,] 1.863693 -0.5328152  0.3626843 0.8205270
#> [11,] 1.350625 -0.7269862  0.1800595 0.2608057
#> [12,] 1.837557 -0.9024051  0.1426004 0.6848011
#> [13,] 1.782969 -0.8813813 -0.3391548 0.8376019
#> [14,] 2.189806 -0.6315769  0.2908370 1.0666460
#> [15,] 2.087206 -0.5232266  0.7920362 0.8567674
#> [16,] 1.903140 -0.4078820  0.7090987 0.7854802
#> [17,] 2.063816 -0.4473501  0.7683450 0.9068884
#> [18,] 1.958371 -0.5659214  0.6917874 0.7067773
#> [19,] 2.120129 -0.3325852  0.6674015 1.0988311
#> [20,] 1.635054 -1.0558912 -0.4952106 0.7797812
#> [21,] 1.703294 -0.8210991  0.1003906 0.6229874
#> [22,] 2.195841 -0.3306081  0.9325142 1.0048743
#> [23,] 2.045415 -0.3693115  0.5729658 1.0012244
#> [24,] 1.368145 -0.8292736  0.1471155 0.5106567
#> [25,] 1.795700 -0.2274408  0.6203214 0.9598149
#> [26,] 1.904037 -0.4347045  0.6131638 0.9210558
#> [27,] 1.022956 -0.9193970 -0.6619793 0.1777182
#> [28,] 2.317422 -0.5955481  0.2986447 1.0490099
#> [29,] 1.817416 -0.4858521  0.4965360 0.8808276
#> [30,] 1.830155 -0.7455899  0.2718785 0.6731312
#> [31,] 1.520079 -1.0967454 -0.2864198 0.5480350
#> [32,] 2.186005 -0.4877537  0.5293647 1.1065203
#> [33,] 1.863961 -0.8684577 -0.1915545 0.8759312
#> [34,] 2.198826 -0.3689436  0.8254421 1.0846275
#> [35,] 2.277357 -0.3220287  1.0862768 0.9392774
#> [36,] 1.688114 -0.5857899  0.4336878 0.7489261
#> [37,] 1.839295 -0.5091448  0.5717386 0.7835727
#> [38,] 1.661179 -0.4222171  0.6766004 0.6706903
#> [39,] 2.092596 -0.5810758  0.7277483 0.8787350
#> [40,] 1.702565 -0.6379722  0.3797385 0.6754216
#> 
#> $pre[[3]]
#>              [,1]       [,2]
#>  [1,] -0.08449809 -0.2201229
#>  [2,] -0.08418336 -0.2276010
#>  [3,] -0.08285303 -0.2324833
#>  [4,] -0.08728107 -0.2292074
#>  [5,] -0.07453704 -0.2044959
#>  [6,] -0.05636942 -0.2275277
#>  [7,] -0.02970951 -0.1870208
#>  [8,] -0.08818303 -0.2263109
#>  [9,] -0.07947902 -0.2234163
#> [10,] -0.07000511 -0.2204231
#> [11,] -0.04470188 -0.2122161
#> [12,] -0.05181365 -0.2229709
#> [13,] -0.04366241 -0.2101933
#> [14,] -0.06857932 -0.2246877
#> [15,] -0.07620267 -0.2325158
#> [16,] -0.07862953 -0.2227715
#> [17,] -0.08124466 -0.2317272
#> [18,] -0.07127060 -0.2260429
#> [19,] -0.08972334 -0.2259049
#> [20,] -0.03755629 -0.2093681
#> [21,] -0.04637628 -0.2206549
#> [22,] -0.09060480 -0.2321276
#> [23,] -0.08404904 -0.2206858
#> [24,] -0.04975848 -0.2149410
#> [25,] -0.08439242 -0.2169564
#> [26,] -0.08063549 -0.2227035
#> [27,] -0.02773564 -0.1842670
#> [28,] -0.06554780 -0.2262761
#> [29,] -0.07292735 -0.2278474
#> [30,] -0.05916633 -0.2256149
#> [31,] -0.03926406 -0.2132845
#> [32,] -0.07621799 -0.2313957
#> [33,] -0.04825670 -0.2165979
#> [34,] -0.08895517 -0.2297553
#> [35,] -0.09035727 -0.2334493
#> [36,] -0.06651827 -0.2242312
#> [37,] -0.07878671 -0.2220241
#> [38,] -0.08009796 -0.2192471
#> [39,] -0.07401255 -0.2345963
#> [40,] -0.06170986 -0.2240731
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5338543  0.5338543
#>  [2,] -0.5357931  0.5357931
#>  [3,]  0.4626621 -0.4626621
#>  [4,] -0.5354221  0.5354221
#>  [5,] -0.5324441  0.5324441
#>  [6,]  0.4573146 -0.4573146
#>  [7,]  0.4607531 -0.4607531
#>  [8,]  0.4655228 -0.4655228
#>  [9,]  0.4640777 -0.4640777
#> [10,]  0.4624662 -0.4624662
#> [11,] -0.5417809  0.5417809
#> [12,]  0.4573148 -0.4573148
#> [13,]  0.4584632 -0.4584632
#> [14,] -0.5389480  0.5389480
#> [15,]  0.4610011 -0.4610011
#> [16,] -0.5359732  0.5359732
#> [17,]  0.4624502 -0.4624502
#> [18,] -0.5386160  0.5386160
#> [19,] -0.5339929  0.5339929
#> [20,]  0.4571524 -0.4571524
#> [21,] -0.5434597  0.5434597
#> [22,]  0.4646782 -0.4646782
#> [23,] -0.5341062  0.5341062
#> [24,]  0.4587980 -0.4587980
#> [25,] -0.5330926  0.5330926
#> [26,]  0.4645426 -0.4645426
#> [27,]  0.4609469 -0.4609469
#> [28,] -0.5400958  0.5400958
#> [29,]  0.4613473 -0.4613473
#> [30,]  0.4584837 -0.4584837
#> [31,]  0.4566043 -0.4566043
#> [32,]  0.4612832 -0.4612832
#> [33,]  0.4580138 -0.4580138
#> [34,] -0.5351420  0.5351420
#> [35,] -0.5357121  0.5357121
#> [36,]  0.4606533 -0.4606533
#> [37,]  0.4642518 -0.4642518
#> [38,]  0.4652687 -0.4652687
#> [39,]  0.4599401 -0.4599401
#> [40,]  0.4594981 -0.4594981
#> 
#> $L
#> [1] 0.6894387 0.6888059 0.7011970 0.7917766 1.1081432 0.6746276
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536