Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11          V12         V13         V14
#>  [1,] -0.2503055 -0.168772942 -0.07506883 -0.176258643 -0.06003977 -0.13703077
#>  [2,] -0.2413210 -0.006337514 -0.11247558 -0.025608004 -0.07643118 -0.03211457
#>  [3,] -0.2304612 -0.255478133 -0.05329743 -0.036889080 -0.19382361 -0.07920442
#>  [4,] -0.3805817 -0.118207117 -0.02136619 -0.047581481 -0.05588042 -0.14624335
#>  [5,] -0.2574007 -0.063706722 -0.11897022 -0.073587318 -0.02207183 -0.17955638
#>  [6,] -0.2740628 -0.169209729 -0.16222816 -0.040628429 -0.10512110 -0.01445194
#>  [7,] -0.2660449 -0.047648613 -0.08975022 -0.215154876 -0.09741141 -0.05934780
#>  [8,] -0.2356005 -0.146954460 -0.22916129 -0.110022985 -0.17932864 -0.10036453
#>  [9,] -0.3161740 -0.059772992 -0.06623609  0.003773482  0.02794064 -0.07357643
#> [10,] -0.2461737  0.005792273 -0.13433620 -0.036609336  0.01920035  0.01169122
#>                V15         V16         V17          V18          V19         V2
#>  [1,] -0.060676125 -0.11653650 -0.09127173 -0.045624916 -0.043222921 -0.2575434
#>  [2,]  0.025144996  0.01945673  0.01092415  0.093624554  0.035513385 -0.2166416
#>  [3,] -0.244098831 -0.10136565 -0.02655540 -0.055709850 -0.087885583 -0.3398264
#>  [4,] -0.101040120  0.02129702  0.02855095  0.012660511 -0.011069251 -0.3627631
#>  [5,] -0.008360837  0.04795566  0.07619544 -0.049322888  0.012763146 -0.3150107
#>  [6,]  0.045590368 -0.02996202 -0.04744208  0.007473164 -0.006636834 -0.1559804
#>  [7,] -0.042923746  0.02501041  0.02853585  0.133318842  0.126445609 -0.2546566
#>  [8,] -0.129604014 -0.13874450 -0.12241902 -0.106751043 -0.122677803 -0.1965371
#>  [9,]  0.042184261  0.03254949  0.03661105 -0.079410590  0.008207018 -0.3553558
#> [10,] -0.045318214 -0.06949736  0.04268561  0.119764186  0.200318631 -0.3342404
#>                V20         V21         V22          V23        V24         V25
#>  [1,] -0.005447580  0.01227511 -0.03289664  0.055679948 0.07075641  0.06884824
#>  [2,]  0.170046530  0.15535206  0.15370016  0.158144580 0.28056923  0.15632397
#>  [3,]  0.029884370 -0.01337939  0.18370960  0.247439380 0.25931638  0.21256783
#>  [4,]  0.007426453  0.05166409  0.17230506  0.175828960 0.09532904  0.14731810
#>  [5,] -0.062563210  0.16018228  0.07639818  0.086034337 0.15835578  0.08937761
#>  [6,] -0.098296114 -0.02128944  0.01067599  0.093627237 0.09091102 -0.04207767
#>  [7,]  0.146430560  0.07357595  0.12504496  0.130137405 0.19158817  0.24413967
#>  [8,]  0.049691921 -0.02176649  0.01841860  0.039320898 0.11314443  0.08074893
#>  [9,] -0.053511880  0.05583100  0.04785503 -0.008057131 0.02807107  0.18513535
#> [10,]  0.142499229  0.10191027  0.07343722  0.098434711 0.12305687  0.21165207
#>              V26        V27        V28          V29         V3         V30
#>  [1,] 0.07428121 0.26365389 0.17799526  0.046086040 -0.2641842  0.02501296
#>  [2,] 0.28429676 0.23699732 0.14117443  0.147207871 -0.2111229  0.10272740
#>  [3,] 0.24563607 0.15517944 0.20644246  0.103918309 -0.3634614 -0.04336807
#>  [4,] 0.12902805 0.09003705 0.21713916  0.192686692 -0.4054702  0.11455836
#>  [5,] 0.04785858 0.20736672 0.10853857  0.077158953 -0.2633983  0.10441875
#>  [6,] 0.05878752 0.15390572 0.15185324  0.003523128 -0.1791463 -0.05289534
#>  [7,] 0.14338092 0.04671038 0.05292638 -0.067178225 -0.2942376 -0.08120792
#>  [8,] 0.16262541 0.03740743 0.01343134  0.165706431 -0.2146769  0.12292760
#>  [9,] 0.11155561 0.08651919 0.04521768  0.131347247 -0.3106859  0.13649177
#> [10,] 0.17476583 0.11358749 0.10960072  0.066358048 -0.2528563 -0.04580903
#>               V31          V32          V33          V34          V35
#>  [1,]  0.03560355  0.034682757 -0.134280974 -0.016576502  0.026165178
#>  [2,]  0.07817520 -0.009897402  0.060130978 -0.030456871  0.101736848
#>  [3,]  0.07479034  0.005313163 -0.033186819 -0.050997411 -0.154016543
#>  [4,]  0.11388768  0.001933935  0.107919620  0.073442467  0.070578915
#>  [5,] -0.04149507  0.002642586  0.012466019  0.018085162 -0.039774055
#>  [6,]  0.05697004 -0.061765750 -0.006526092 -0.001387524 -0.066752735
#>  [7,] -0.02914285 -0.036331334 -0.005016609 -0.073812196  0.085619952
#>  [8,]  0.07017993  0.006029439 -0.013495691  0.043796181  0.001345572
#>  [9,]  0.08818447  0.054836019 -0.085529886 -0.047719979 -0.063219112
#> [10,]  0.07710635 -0.066512311 -0.054887002 -0.049347577 -0.052958217
#>               V36         V37         V38          V39         V4          V40
#>  [1,]  0.07692295 -0.08127354 -0.06704146 -0.049659932 -0.2927481 -0.125587615
#>  [2,] -0.03342176  0.07319368  0.09531079  0.049536576 -0.3098064 -0.071126316
#>  [3,] -0.04796378 -0.11581096 -0.16437079 -0.052637062 -0.1999344 -0.042817124
#>  [4,]  0.14944956  0.09370058  0.05575479  0.002894846 -0.3586013  0.003045973
#>  [5,] -0.07947906  0.02453372 -0.04167147 -0.095562047 -0.2205093 -0.001388846
#>  [6,] -0.02290207 -0.05692508 -0.11821802 -0.054040971 -0.2480342 -0.168163410
#>  [7,]  0.02755842  0.01949115  0.04967652 -0.073523946 -0.1651889 -0.075666323
#>  [8,] -0.04586999  0.01922141 -0.08419395 -0.141529947 -0.2804657 -0.102952719
#>  [9,] -0.06082917  0.02569214 -0.09965198 -0.077820242 -0.1993408 -0.053062155
#> [10,]  0.05955229 -0.03972719  0.02062620 -0.108236341 -0.3261822  0.042426662
#>                V41          V42           V43         V44         V45
#>  [1,] -0.119974752 -0.033477056 -0.1793998793 -0.05119311 -0.07376470
#>  [2,] -0.094145457 -0.006240064  0.0003850696 -0.11707399 -0.18292440
#>  [3,] -0.146250910 -0.153458137 -0.0194121691 -0.17295760 -0.10303926
#>  [4,]  0.054110502  0.047093847 -0.0415999005 -0.15674466 -0.04410103
#>  [5,] -0.081296113 -0.039884898 -0.0824367275 -0.06729974 -0.14304528
#>  [6,] -0.124134454 -0.066810410 -0.1070687549 -0.06947136 -0.19539539
#>  [7,] -0.019087363 -0.072256960  0.0241964165 -0.01470052 -0.14513089
#>  [8,] -0.078394306 -0.199738407 -0.0848632231 -0.17016459 -0.13881182
#>  [9,] -0.091036532 -0.106006999 -0.1443167888 -0.05003129 -0.06418207
#> [10,] -0.005141947 -0.132833046 -0.1142065906 -0.10679671 -0.16449936
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.22846417 -0.28936317 -0.1812415 -0.2077723 -0.2435716 -0.3330383
#>  [2,] -0.24396349 -0.20379900 -0.2532232 -0.2309042 -0.2114445 -0.3146374
#>  [3,] -0.20265994 -0.18585228 -0.1429341 -0.2044920 -0.1724933 -0.3937644
#>  [4,] -0.16454956 -0.20185001 -0.3189616 -0.2802751 -0.2709102 -0.3183506
#>  [5,] -0.11030763 -0.27396383 -0.2138984 -0.3030237 -0.2967975 -0.2662510
#>  [6,] -0.05638310 -0.24100567 -0.2472341 -0.1686555 -0.1833712 -0.3532257
#>  [7,] -0.08754195 -0.17707058 -0.3036521 -0.3538162 -0.2453106 -0.2916574
#>  [8,] -0.15119104 -0.10264432 -0.1730504 -0.3363955 -0.2322272 -0.3373919
#>  [9,] -0.16279947 -0.09903966 -0.2710420 -0.2953398 -0.2650756 -0.2864227
#> [10,] -0.14187663 -0.18710191 -0.1387240 -0.2736824 -0.2924591 -0.3472577
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2719214 -0.2677735 -0.3669426 -0.2569837 -0.3658469 -0.2878231
#>  [2,] -0.3526579 -0.2643547 -0.2160604 -0.3466251 -0.1969018 -0.2688206
#>  [3,] -0.3601658 -0.3701232 -0.2768590 -0.3607542 -0.2264904 -0.2612774
#>  [4,] -0.3262187 -0.3192143 -0.4305146 -0.3511585 -0.4569923 -0.3897793
#>  [5,] -0.3078527 -0.3094656 -0.4014901 -0.3309051 -0.2772637 -0.4198945
#>  [6,] -0.3948152 -0.3501884 -0.2767626 -0.2906595 -0.3357563 -0.3721266
#>  [7,] -0.2351406 -0.2598154 -0.4055521 -0.3517436 -0.3712569 -0.3504529
#>  [8,] -0.3181653 -0.3954600 -0.2909521 -0.3752750 -0.2728457 -0.3052524
#>  [9,] -0.3326142 -0.3631081 -0.3826615 -0.2566837 -0.4060863 -0.4249363
#> [10,] -0.3799965 -0.2430898 -0.2585260 -0.3742194 -0.3825728 -0.3280978
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.3985695 -0.3312482 -0.2692405 -0.2665359 -0.3694376 -0.10815017
#>  [2,] -0.2734888 -0.2096142 -0.3603471 -0.1210813 -0.2705554 -0.09214161
#>  [3,] -0.3621285 -0.3233546 -0.3787034 -0.1563673 -0.3987046 -0.22287005
#>  [4,] -0.3588782 -0.4383667 -0.3566156 -0.3035028 -0.3244143 -0.27848611
#>  [5,] -0.3856759 -0.3402591 -0.2542027 -0.2020596 -0.2912622 -0.16069724
#>  [6,] -0.2366724 -0.2730389 -0.3864813 -0.1734643 -0.2760362 -0.24200283
#>  [7,] -0.2870899 -0.3643506 -0.3584102 -0.1088468 -0.3896527 -0.22473961
#>  [8,] -0.3929006 -0.4132891 -0.2831670 -0.2070950 -0.3788642 -0.16907605
#>  [9,] -0.3725989 -0.2984704 -0.3924783 -0.2052812 -0.3442452 -0.08724009
#> [10,] -0.3610846 -0.2608023 -0.2303032 -0.1069351 -0.3025696 -0.16587347
#>                V8          V9
#>  [1,] -0.15590387 -0.21319615
#>  [2,] -0.14219231 -0.16823244
#>  [3,] -0.12729498 -0.18478659
#>  [4,] -0.25117791 -0.17844242
#>  [5,] -0.14703142 -0.18035974
#>  [6,] -0.15491877 -0.18182965
#>  [7,] -0.22428634 -0.03533377
#>  [8,] -0.24039800 -0.08293137
#>  [9,] -0.14153803 -0.05102603
#> [10,] -0.04961377 -0.04508368
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,] -0.12923227 -0.03489598 -0.103056486 -0.05086461 -0.12757841 -0.02061523
#> [2,]  0.03394859  0.15301241  0.003654258  0.11586145 -0.01741888  0.06368982
#>             [,7]        [,8]         [,9]         [,10]
#> [1,] -0.05844676 -0.02214842 -0.025031464 -0.0007314676
#> [2,]  0.08498604  0.15909161 -0.001804094  0.0437334058
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.479409e-05  5.721252e-05  1.407906e-04  1.460975e-04  9.863714e-05
#>  [2,]  3.130073e-05  1.721137e-04  2.627562e-04  2.576111e-04  2.140406e-04
#>  [3,]  1.327529e-05  9.890667e-05  1.039972e-04  8.600860e-05  1.031992e-04
#>  [4,]  1.237333e-05  6.369176e-05  8.988132e-05  7.918190e-05  8.244809e-05
#>  [5,]  7.278372e-06  5.244756e-05  3.156958e-05  6.116248e-06  4.644862e-05
#>  [6,] -9.833973e-06 -9.113402e-05 -1.771158e-04 -2.226713e-04 -1.246787e-04
#>  [7,]  1.752697e-05  1.094063e-04  1.466011e-04  1.323668e-04  1.262699e-04
#>  [8,]  1.525850e-05  5.380359e-05  2.084108e-04  2.339895e-04  1.040090e-04
#>  [9,] -3.751696e-05 -1.937020e-04 -4.805227e-04 -5.758984e-04 -3.297337e-04
#> [10,] -2.325309e-05 -1.034990e-04 -3.043679e-04 -3.633031e-04 -1.810850e-04
#>                 V14           V15           V16           V17           V18
#>  [1,]  3.078284e-05 -6.457376e-05 -1.199075e-04 -8.685141e-05 -1.200032e-04
#>  [2,]  1.489380e-04  1.110651e-04  8.947852e-05  1.188845e-04  7.257213e-05
#>  [3,]  1.351653e-04  1.438708e-04  2.115741e-04  2.485376e-04  2.686516e-04
#>  [4,]  9.790993e-05  7.236625e-05  1.006148e-04  1.404489e-04  1.320187e-04
#>  [5,]  1.199316e-04  1.359796e-04  2.199988e-04  2.593534e-04  2.875406e-04
#>  [6,]  5.523970e-05  9.163080e-05  2.026239e-04  2.425814e-04  2.744065e-04
#>  [7,]  1.207798e-04  9.631470e-05  1.208527e-04  1.564232e-04  1.541183e-04
#>  [8,] -8.714030e-05 -2.668225e-04 -4.405608e-04 -4.280285e-04 -4.962363e-04
#>  [9,]  1.098062e-04  3.068168e-04  6.505971e-04  7.023873e-04  8.584128e-04
#> [10,]  1.198277e-04  2.676249e-04  5.313619e-04  5.705930e-04  6.976061e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0000730030  2.501237e-05 -1.130586e-05  1.793883e-05  1.113098e-04
#>  [2,]  0.0001531736  3.752513e-05  2.696859e-04  3.081410e-04  3.185664e-04
#>  [3,]  0.0002865410  1.931761e-05  2.865092e-04  2.975671e-04  3.186417e-04
#>  [4,]  0.0001388232  2.113198e-05  1.379869e-04  1.491377e-04  1.982651e-04
#>  [5,]  0.0002742181  1.415304e-05  2.339921e-04  2.344120e-04  2.687307e-04
#>  [6,]  0.0001768776 -3.640558e-06  5.099332e-05  2.513381e-05  8.123204e-05
#>  [7,]  0.0001929259  2.484546e-05  2.253594e-04  2.469255e-04  2.889030e-04
#>  [8,] -0.0004044246  2.553091e-05 -2.664094e-04 -2.177952e-04 -8.381405e-05
#>  [9,]  0.0006205838 -3.316023e-05  2.512551e-04  1.651332e-04  1.772380e-04
#> [10,]  0.0005454421 -1.866101e-05  2.745347e-04  2.060230e-04  2.095220e-04
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0001067540 -2.259255e-06 -1.067125e-04 -2.712769e-04 -3.209027e-04
#>  [2,]  0.0002515693  1.492672e-04  6.008116e-05 -4.049785e-05 -4.043534e-05
#>  [3,]  0.0003497607  3.421201e-04  3.483712e-04  3.646251e-04  3.299593e-04
#>  [4,]  0.0002248710  1.777597e-04  1.488596e-04  9.907261e-05  4.897279e-05
#>  [5,]  0.0003355823  3.498914e-04  3.831856e-04  4.147929e-04  3.537054e-04
#>  [6,]  0.0002215325  2.963285e-04  3.958329e-04  4.701630e-04  3.661970e-04
#>  [7,]  0.0003050913  2.581326e-04  2.241048e-04  1.782599e-04  1.385425e-04
#>  [8,] -0.0001414673 -3.341197e-04 -5.379593e-04 -8.457699e-04 -8.833959e-04
#>  [9,]  0.0004972848  7.821323e-04  1.120231e-03  1.496207e-03  1.332833e-03
#> [10,]  0.0004450859  6.562588e-04  9.089284e-04  1.211258e-03  1.088223e-03
#>                 V28           V29            V3           V30           V31
#>  [1,] -2.570343e-04 -1.730893e-04  7.939219e-06 -1.013163e-04 -0.0001525937
#>  [2,] -1.376981e-06  2.011104e-05  2.713509e-05  1.171118e-05 -0.0001013425
#>  [3,]  3.087054e-04  2.675758e-04  2.235796e-05  2.717969e-04  0.0002582357
#>  [4,]  5.938784e-05  6.475365e-05  1.711513e-05  1.162149e-04  0.0001184788
#>  [5,]  3.212191e-04  2.764555e-04  1.902235e-05  3.098340e-04  0.0003483074
#>  [6,]  2.999366e-04  2.463246e-04  5.875111e-06  3.446459e-04  0.0005146736
#>  [7,]  1.480244e-04  1.476058e-04  2.098893e-05  1.646906e-04  0.0001250953
#>  [8,] -7.420180e-04 -5.474731e-04 -7.433063e-06 -4.494778e-04 -0.0005831363
#>  [9,]  1.101286e-03  8.489468e-04  1.038125e-05  9.442675e-04  0.0013514378
#> [10,]  9.382266e-04  7.655068e-04  1.255157e-05  8.280108e-04  0.0010812332
#>                 V32           V33           V34           V35           V36
#>  [1,] -4.912216e-05 -4.411978e-05 -0.0001611450 -2.928540e-04 -3.904936e-04
#>  [2,] -2.789823e-06 -1.538528e-05 -0.0001072379 -2.475913e-04 -4.132894e-04
#>  [3,]  2.464681e-04  2.629030e-04  0.0002575495  2.971404e-04  3.072358e-04
#>  [4,]  1.487423e-04  1.738413e-04  0.0001316388  1.224448e-04  1.048983e-04
#>  [5,]  3.163235e-04  3.438455e-04  0.0003596667  4.485082e-04  5.014822e-04
#>  [6,]  4.371620e-04  4.857491e-04  0.0005630451  7.633259e-04  9.045634e-04
#>  [7,]  1.570973e-04  1.648064e-04  0.0001153724  8.346729e-05  3.654399e-05
#>  [8,] -3.865444e-04 -4.058927e-04 -0.0006168660 -9.178184e-04 -1.110373e-03
#>  [9,]  1.049043e-03  1.152423e-03  0.0014450834  2.051783e-03  2.525889e-03
#> [10,]  8.330607e-04  9.253557e-04  0.0011483334  1.612204e-03  1.980172e-03
#>                 V37           V38           V39            V4           V40
#>  [1,] -3.752664e-04 -0.0001765069 -1.365121e-04  4.446370e-05 -1.910728e-04
#>  [2,] -3.802651e-04 -0.0001593729 -1.065171e-04  6.044765e-05 -1.875285e-04
#>  [3,]  2.887679e-04  0.0002359378  1.835575e-04  3.602280e-05  1.955083e-04
#>  [4,]  9.220520e-05  0.0001138199  8.296280e-05  3.979911e-05  7.823383e-05
#>  [5,]  4.576731e-04  0.0003299978  2.414471e-04  2.867321e-05  2.800909e-04
#>  [6,]  7.918011e-04  0.0004852116  3.131295e-04  1.259758e-06  4.081214e-04
#>  [7,]  3.685057e-05  0.0000862196  7.238393e-05  4.491897e-05  5.408849e-05
#>  [8,] -1.034475e-03 -0.0005979757 -4.491864e-04  4.368186e-05 -5.602993e-04
#>  [9,]  2.312491e-03  0.0013950857  9.799316e-04 -4.091119e-05  1.275833e-03
#> [10,]  1.821186e-03  0.0011388113  8.229394e-04 -2.478111e-05  1.026252e-03
#>                 V41          V42           V43           V44           V45
#>  [1,] -4.177522e-05 9.609597e-05  1.027857e-04  1.323881e-04  2.172297e-04
#>  [2,] -2.481929e-05 1.366249e-04  1.494699e-04  1.784585e-04  2.774868e-04
#>  [3,]  1.735402e-04 1.352186e-04  1.061101e-04  6.902491e-05  5.419025e-05
#>  [4,]  1.156130e-04 1.278099e-04  1.007660e-04  7.645220e-05  1.015007e-04
#>  [5,]  2.220011e-04 1.306055e-04  8.498440e-05  2.912590e-05 -5.277810e-06
#>  [6,]  2.741941e-04 7.256988e-05 -1.876623e-06 -8.635231e-05 -1.623268e-04
#>  [7,]  1.034299e-04 1.383427e-04  1.218472e-04  1.081230e-04  1.399369e-04
#>  [8,] -2.673574e-04 2.883351e-05  8.155402e-05  1.867391e-04  3.478542e-04
#>  [9,]  7.558671e-04 1.061917e-04 -7.931464e-05 -3.233343e-04 -6.236587e-04
#> [10,]  6.001665e-04 1.109794e-04 -1.282374e-05 -2.159926e-04 -4.667570e-04
#>                 V46           V47           V48           V49           V5
#>  [1,]  1.861757e-04  1.175862e-04  6.334722e-05  3.685755e-05 4.081137e-05
#>  [2,]  2.368371e-04  1.444795e-04  9.227820e-05  5.142462e-05 5.337659e-05
#>  [3,]  5.399899e-05  3.164656e-05  3.109560e-05  2.109685e-05 4.941251e-05
#>  [4,]  9.970714e-05  6.329123e-05  4.387280e-05  2.744175e-05 4.734698e-05
#>  [5,]  1.344904e-05  6.989108e-06  1.510950e-05  1.233107e-05 4.519535e-05
#>  [6,] -1.030698e-04 -6.802092e-05 -3.461208e-05 -1.812409e-05 1.977521e-05
#>  [7,]  1.272765e-04  7.793010e-05  5.344487e-05  3.260277e-05 5.081413e-05
#>  [8,]  2.813204e-04  1.809822e-04  8.491460e-05  4.770081e-05 2.175781e-05
#>  [9,] -4.534938e-04 -2.824234e-04 -1.406915e-04 -7.121585e-05 1.529789e-05
#> [10,] -3.451248e-04 -2.173347e-04 -1.037620e-04 -4.838278e-05 2.672197e-05
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.339927e-05  2.158823e-05  1.644851e-05  2.824102e-06  2.097712e-06
#>  [2,]  1.672666e-05  2.726693e-05  2.065390e-05  5.121477e-06  6.537700e-06
#>  [3,]  1.002833e-05  5.966165e-06  6.412205e-06  5.936173e-06  4.359196e-06
#>  [4,]  1.222429e-05  1.194295e-05  1.031518e-05  5.210089e-06  3.570123e-06
#>  [5,]  9.274753e-06  1.415483e-06  3.548357e-06  6.180794e-06  3.422624e-06
#>  [6,]  3.693723e-06 -1.106368e-05 -4.599641e-06  5.404011e-06  6.646287e-07
#>  [7,]  1.338956e-05  1.445531e-05  1.191918e-05  5.496465e-06  4.417255e-06
#>  [8,]  1.226205e-05  3.201918e-05  2.148836e-05 -2.231674e-06 -1.361943e-06
#>  [9,] -6.046246e-06 -5.128879e-05 -3.137851e-05  9.620748e-06 -1.068137e-06
#> [10,] -3.258808e-06 -3.918077e-05 -2.367600e-05  8.773756e-06  3.671956e-07
#>                 V55           V56           V57           V58          V59
#>  [1,]  1.572208e-06  2.359782e-06  8.651260e-07  2.986483e-06 4.646937e-06
#>  [2,]  5.207756e-06  5.127930e-06  2.732822e-06  4.063429e-06 5.632572e-06
#>  [3,]  4.667927e-06  4.814452e-06  3.280119e-06  4.619663e-06 4.340160e-06
#>  [4,]  4.333126e-06  3.704059e-06  2.789335e-06  4.595114e-06 5.137737e-06
#>  [5,]  4.858540e-06  4.448342e-06  3.526623e-06  4.826494e-06 4.130931e-06
#>  [6,]  4.506379e-06  2.213904e-06  3.544260e-06  3.733784e-06 2.594244e-06
#>  [7,]  4.505130e-06  4.664404e-06  2.826486e-06  4.534115e-06 4.896755e-06
#>  [8,] -3.539360e-06 -5.651964e-07 -2.512747e-06 -5.297625e-08 2.780267e-06
#>  [9,]  7.731743e-06  3.359600e-06  6.709860e-06  6.384862e-06 4.980106e-07
#> [10,]  5.792297e-06  4.265978e-06  5.619016e-06  6.475290e-06 2.031554e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  5.598235e-08  1.727590e-06  3.329857e-07  3.704152e-05  4.763361e-05
#>  [2,]  2.816819e-05  2.824499e-06  2.397257e-05  7.327786e-05  1.460852e-04
#>  [3,]  5.404189e-05  4.211961e-06  7.134251e-05  8.035916e-05  7.250765e-05
#>  [4,]  3.476324e-05  3.667978e-06  4.822133e-05  6.441479e-05  4.355033e-05
#>  [5,]  5.569357e-05  4.636330e-06  7.966179e-05  7.683717e-05  3.178223e-05
#>  [6,]  4.272141e-05  4.319174e-06  7.387547e-05  3.759636e-05 -8.455286e-05
#>  [7,]  4.022694e-05  3.816165e-06  5.279417e-05  7.778015e-05  8.266806e-05
#>  [8,] -5.642513e-05 -1.673836e-06 -7.328195e-05 -3.703892e-06  5.446234e-05
#>  [9,]  1.177094e-04  8.954624e-06  1.991726e-04  8.291616e-05 -2.013900e-04
#> [10,]  1.072267e-04  7.762955e-06  1.747139e-04  8.880106e-05 -1.231971e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#> [1,]  0.06302599  0.07348821  0.0559923  0.0757229  0.05976218  0.05118534
#> [2,] -0.06302599 -0.07348821 -0.0559923 -0.0757229 -0.05976218 -0.05118534
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.06574474  0.05651553  0.06074106  0.06381055
#> [2,] -0.06574474 -0.05651553 -0.06074106 -0.06381055
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5818934 0.5901088 0.4160727 0.6132988 0.3858137 0.3090496 0.5792395
#>  [8] 0.3496410 0.6176813 0.4227413
#> 
#> $B[[2]]
#> [1] -0.04143934  0.06583914
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -2.444753e-04  3.535846e-06  5.656337e-04  2.789315e-04  6.667523e-04
#>  [6]  7.838820e-04  3.221199e-04 -1.063241e-03  2.206468e-03  1.815858e-03
#> 
#> $vB[[2]]
#> [1]  0.08426905 -0.08426905
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#>  [2,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [3,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#>  [4,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#>  [5,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#>  [6,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#>  [7,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [8,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#>  [9,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [10,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [11,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [12,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [13,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [14,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [15,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [16,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [17,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [18,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [19,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [20,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [21,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [22,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [23,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [24,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [25,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [26,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [27,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [28,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [29,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [30,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [31,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [32,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [33,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [34,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [35,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [36,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [37,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [38,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [39,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [40,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#>  [2,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [3,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#>  [4,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#>  [5,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#>  [6,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#>  [7,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [8,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#>  [9,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [10,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [11,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [12,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [13,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [14,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [15,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [16,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [17,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [18,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [19,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [20,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [21,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [22,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [23,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [24,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [25,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [26,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [27,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [28,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [29,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [30,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [31,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [32,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [33,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [34,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [35,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [36,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [37,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [38,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [39,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [40,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#>  [2,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [3,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#>  [4,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#>  [5,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#>  [6,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#>  [7,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [8,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#>  [9,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [10,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [11,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [12,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [13,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [14,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [15,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [16,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [17,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [18,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [19,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [20,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [21,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [22,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [23,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [24,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [25,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [26,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [27,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [28,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [29,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [30,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [31,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [32,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [33,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [34,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [35,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [36,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [37,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [38,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [39,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [40,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#>  [2,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [3,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#>  [4,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#>  [5,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#>  [6,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#>  [7,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [8,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#>  [9,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [10,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [11,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [12,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [13,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [14,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [15,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [16,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [17,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [18,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [19,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [20,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [21,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [22,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [23,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [24,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [25,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [26,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [27,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [28,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [29,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [30,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [31,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [32,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [33,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [34,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [35,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [36,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [37,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [38,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [39,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [40,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#>  [2,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [3,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#>  [4,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#>  [5,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#>  [6,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#>  [7,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [8,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#>  [9,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [10,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [11,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [12,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [13,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [14,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [15,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [16,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [17,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [18,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [19,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [20,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [21,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [22,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [23,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [24,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [25,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [26,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [27,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [28,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [29,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [30,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [31,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [32,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [33,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [34,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [35,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [36,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [37,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [38,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [39,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [40,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#>  [2,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [3,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#>  [4,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#>  [5,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#>  [6,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#>  [7,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [8,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#>  [9,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [10,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [11,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [12,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [13,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [14,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [15,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [16,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [17,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [18,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [19,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [20,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [21,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [22,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [23,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [24,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [25,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [26,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [27,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [28,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [29,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [30,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [31,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [32,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [33,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [34,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [35,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [36,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [37,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [38,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [39,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [40,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5170780 0.7781554 0.5159256 0.7858504 0.5322100 0.3746640 0.5855862
#>  [2,] 0.5383683 0.8444671 0.6012551 0.7861895 0.6248722 0.4582326 0.7357660
#>  [3,] 0.6718564 0.8295881 0.7120502 0.7798854 0.6629843 0.5564217 0.7179210
#>  [4,] 0.6457688 0.8544624 0.6649229 0.8177124 0.6650264 0.5263537 0.7464923
#>  [5,] 0.5819894 0.7722406 0.5607772 0.7709990 0.5957371 0.4795775 0.6230316
#>  [6,] 0.4993662 0.8515412 0.5661992 0.7644626 0.6075684 0.4366008 0.7620161
#>  [7,] 0.6785993 0.8876878 0.7706597 0.8354338 0.7103932 0.5543474 0.7627080
#>  [8,] 0.5492035 0.7526793 0.4560126 0.7917536 0.5476710 0.4251750 0.5793373
#>  [9,] 0.5879264 0.8373594 0.5887030 0.7740263 0.6307564 0.4920699 0.7179632
#> [10,] 0.6049613 0.8647238 0.6987581 0.8250629 0.6528935 0.4862941 0.7559370
#> [11,] 0.5150187 0.7866547 0.5395173 0.7712333 0.5559979 0.3952125 0.6165154
#> [12,] 0.5859637 0.8553441 0.6813244 0.7789867 0.6584280 0.5122263 0.7909893
#> [13,] 0.4376760 0.7426235 0.4211706 0.6426354 0.5272223 0.4096036 0.6585644
#> [14,] 0.5982322 0.8239689 0.6168859 0.7950227 0.6201576 0.4744628 0.6903689
#> [15,] 0.5362898 0.8102569 0.5284234 0.7159857 0.5724173 0.4339293 0.6212523
#> [16,] 0.5262783 0.8666378 0.6091905 0.7868872 0.6525137 0.4743081 0.7381911
#> [17,] 0.6465080 0.8527826 0.6983340 0.7979147 0.6830942 0.5490383 0.7631808
#> [18,] 0.4121741 0.6674649 0.3973914 0.6581443 0.4411072 0.3305823 0.5144762
#> [19,] 0.6293953 0.8417342 0.6862689 0.7928003 0.6512017 0.5116598 0.6796993
#> [20,] 0.5865451 0.8587402 0.6924402 0.8000801 0.6555237 0.4857743 0.7830489
#> [21,] 0.5787579 0.8525782 0.6544323 0.7670474 0.6486552 0.4993012 0.7726578
#> [22,] 0.4010175 0.7552037 0.4286475 0.6232411 0.5271895 0.4044788 0.6581681
#> [23,] 0.5248068 0.7444759 0.5343920 0.6908377 0.5679733 0.4620843 0.6153745
#> [24,] 0.5453775 0.8382802 0.6034832 0.7704264 0.6408252 0.4806574 0.7489009
#> [25,] 0.5862243 0.8563648 0.6448633 0.7465529 0.6391173 0.4950707 0.7492610
#> [26,] 0.5711795 0.7538702 0.4755660 0.7986859 0.5603723 0.4390262 0.5949301
#> [27,] 0.6336118 0.9098207 0.7425116 0.8269893 0.7284064 0.5557190 0.8286796
#> [28,] 0.5315047 0.8460469 0.5918698 0.7864692 0.6232940 0.4492980 0.7573732
#> [29,] 0.6455496 0.8864696 0.6870220 0.8584224 0.6912070 0.5236214 0.7470335
#> [30,] 0.6378856 0.9121075 0.7461269 0.8276772 0.7354696 0.5604725 0.8282880
#> [31,] 0.4969512 0.8515855 0.5574226 0.7842364 0.6070503 0.4109935 0.7631819
#> [32,] 0.6655038 0.8759318 0.7075294 0.8239937 0.6864668 0.5448473 0.7565262
#> [33,] 0.6414506 0.8449243 0.6345835 0.8308954 0.6579046 0.5148343 0.6884690
#> [34,] 0.5805424 0.8008597 0.5266417 0.8042917 0.5672154 0.4387283 0.6401429
#> [35,] 0.4778676 0.7721598 0.5219606 0.7668040 0.5388468 0.3747135 0.5925010
#> [36,] 0.6449044 0.8796888 0.7542142 0.8101858 0.6940235 0.5351364 0.7444395
#> [37,] 0.4354682 0.7615024 0.3611050 0.7677863 0.5002858 0.3347031 0.5981832
#> [38,] 0.5389383 0.8507080 0.6308952 0.7867078 0.6104758 0.4174238 0.6651392
#> [39,] 0.5161294 0.8534061 0.5980480 0.7807826 0.6169289 0.4407778 0.7657658
#> [40,] 0.5996254 0.8899131 0.7171662 0.8039950 0.6865340 0.5105019 0.7756647
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4427414 0.5941526 0.5871591
#>  [2,] 0.4276685 0.6485569 0.7229687
#>  [3,] 0.5859244 0.7037238 0.7176256
#>  [4,] 0.5382075 0.6759466 0.7291759
#>  [5,] 0.5202687 0.6278168 0.6385072
#>  [6,] 0.4033074 0.6151860 0.7325102
#>  [7,] 0.6064995 0.7427635 0.7675301
#>  [8,] 0.4583476 0.5659284 0.5683073
#>  [9,] 0.4714946 0.6559789 0.7339486
#> [10,] 0.5320552 0.6597989 0.7420957
#> [11,] 0.4361236 0.6018463 0.5905652
#> [12,] 0.5280424 0.6697437 0.7699263
#> [13,] 0.3369359 0.5668624 0.6494518
#> [14,] 0.4750800 0.6575872 0.6904211
#> [15,] 0.4352154 0.6703599 0.6719032
#> [16,] 0.4579276 0.6907045 0.7570184
#> [17,] 0.5302247 0.7046819 0.7587809
#> [18,] 0.3312165 0.5023698 0.4886431
#> [19,] 0.5445619 0.6986080 0.7163929
#> [20,] 0.5068059 0.6587571 0.7507422
#> [21,] 0.4723585 0.6562980 0.7612804
#> [22,] 0.2855747 0.5668427 0.6543794
#> [23,] 0.4249952 0.6066950 0.6153462
#> [24,] 0.4728793 0.6548566 0.7329415
#> [25,] 0.4690672 0.6463176 0.7350843
#> [26,] 0.4874832 0.5841355 0.5814685
#> [27,] 0.5236929 0.7201030 0.8139542
#> [28,] 0.4401644 0.6179355 0.7275265
#> [29,] 0.5527223 0.7097507 0.7385298
#> [30,] 0.5268692 0.7276794 0.8177596
#> [31,] 0.3838487 0.6096663 0.7353120
#> [32,] 0.5683875 0.7006248 0.7663175
#> [33,] 0.5689200 0.6891429 0.6903917
#> [34,] 0.5030945 0.6235284 0.6145373
#> [35,] 0.4064570 0.5805622 0.5808207
#> [36,] 0.5633808 0.7229805 0.7644636
#> [37,] 0.3492987 0.5471856 0.5934337
#> [38,] 0.4125302 0.6470806 0.6958163
#> [39,] 0.4266756 0.6165950 0.7309892
#> [40,] 0.4943322 0.7063770 0.7852403
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5206599 0.4793401
#>  [2,] 0.5244499 0.4755501
#>  [3,] 0.5226686 0.4773314
#>  [4,] 0.5226048 0.4773952
#>  [5,] 0.5213157 0.4786843
#>  [6,] 0.5243540 0.4756460
#>  [7,] 0.5234882 0.4765118
#>  [8,] 0.5194545 0.4805455
#>  [9,] 0.5239766 0.4760234
#> [10,] 0.5226344 0.4773656
#> [11,] 0.5211070 0.4788930
#> [12,] 0.5239735 0.4760265
#> [13,] 0.5245318 0.4754682
#> [14,] 0.5230286 0.4769714
#> [15,] 0.5244460 0.4755540
#> [16,] 0.5256971 0.4743029
#> [17,] 0.5242835 0.4757165
#> [18,] 0.5204973 0.4795027
#> [19,] 0.5232922 0.4767078
#> [20,] 0.5234595 0.4765405
#> [21,] 0.5244907 0.4755093
#> [22,] 0.5257799 0.4742201
#> [23,] 0.5229021 0.4770979
#> [24,] 0.5242024 0.4757976
#> [25,] 0.5237448 0.4762552
#> [26,] 0.5195072 0.4804928
#> [27,] 0.5253420 0.4746580
#> [28,] 0.5235071 0.4764929
#> [29,] 0.5229768 0.4770232
#> [30,] 0.5255308 0.4744692
#> [31,] 0.5244228 0.4755772
#> [32,] 0.5232003 0.4767997
#> [33,] 0.5217525 0.4782475
#> [34,] 0.5202528 0.4797472
#> [35,] 0.5213130 0.4786870
#> [36,] 0.5240713 0.4759287
#> [37,] 0.5217281 0.4782719
#> [38,] 0.5240690 0.4759310
#> [39,] 0.5237943 0.4762057
#> [40,] 0.5253382 0.4746618
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]      [,2]        [,3]      [,4]         [,5]         [,6]
#>  [1,]  0.068338643 1.2549491  0.06372391 1.3000918  0.129018753 -0.512259619
#>  [2,]  0.153775690 1.6918482  0.41069746 1.3021077  0.510280225 -0.167459804
#>  [3,]  0.716593230 1.5827110  0.90536220 1.2649985  0.676621777  0.226651969
#>  [4,]  0.600491600 1.7700380  0.68531027 1.5009253  0.685774918  0.105512447
#>  [5,]  0.330945532 1.2210061  0.24431695 1.2139609  0.387734090 -0.081735473
#>  [6,] -0.002535198 1.7467401  0.26636041 1.1773036  0.437102425 -0.254969062
#>  [7,]  0.747342347 2.0673375  1.21204018 1.6246379  0.897294548  0.218251687
#>  [8,]  0.197452993 1.1129535 -0.17640570 1.3355284  0.191264903 -0.301564915
#>  [9,]  0.355399827 1.6387108  0.35860642 1.2311875  0.535463068 -0.031723068
#> [10,]  0.426180841 1.8550914  0.84139093 1.5510332  0.631782340 -0.054837450
#> [11,]  0.060092990 1.3048772  0.15839944 1.2152881  0.224935068 -0.425453546
#> [12,]  0.347304116 1.7771459  0.75986469 1.2597709  0.656296535  0.048915087
#> [13,] -0.250599141 1.0596494 -0.31796998 0.5868212  0.108997023 -0.365604147
#> [14,]  0.398104708 1.5434720  0.47635102 1.3554716  0.490217325 -0.102237589
#> [15,]  0.145415010 1.4516801  0.11381635 0.9246354  0.291720378 -0.265837382
#> [16,]  0.105210059 1.8715527  0.44391068 1.3062635  0.630106839 -0.102858002
#> [17,]  0.603724538 1.7565939  0.83937696 1.3733119  0.768028266  0.196785711
#> [18,] -0.354985088 0.6967412 -0.41634609 0.6550355 -0.236669950 -0.705552446
#> [19,]  0.529623331 1.6711881  0.78273322 1.3418883  0.624325494  0.046647577
#> [20,]  0.349701050 1.8048659  0.81155232 1.3867949  0.643409200 -0.056918034
#> [21,]  0.317676671 1.7549670  0.63857936 1.1917138  0.613133345 -0.002795357
#> [22,] -0.401227204 1.1265610 -0.28737158 0.5033278  0.108865242 -0.386837812
#> [23,]  0.099308544 1.0693636  0.13778562 0.8040383  0.273587100 -0.151954709
#> [24,]  0.182010962 1.6454874  0.41999960 1.2107206  0.578947572 -0.077408931
#> [25,]  0.348378542 1.7854198  0.59653546 1.0803116  0.571535241 -0.019717687
#> [26,]  0.286664979 1.1193613 -0.09781383 1.3781015  0.242673300 -0.245115235
#> [27,]  0.547742878 2.3114471  1.05906382 1.5644385  0.986552250  0.223805688
#> [28,]  0.126185867 1.7039266  0.37170031 1.3037723  0.503553063 -0.203507563
#> [29,]  0.599533703 2.0551767  0.78623349 1.8022483  0.805768419  0.094555936
#> [30,]  0.566198689 2.3396438  1.07806103 1.5692534  1.022552819  0.243079923
#> [31,] -0.012195298 1.7470912  0.23070812 1.2905269  0.434930189 -0.359859762
#> [32,]  0.687918923 1.9544565  0.88341513 1.5436432  0.783652443  0.179872501
#> [33,]  0.581665574 1.6953335  0.55193125 1.5919863  0.653970416  0.059354817
#> [34,]  0.325000456 1.3916763  0.10666768 1.4133365  0.270498800 -0.246324776
#> [35,] -0.088587457 1.2205471  0.08789895 1.1903521  0.155700896 -0.512048331
#> [36,]  0.596714930 1.9894866  1.12121603 1.4512180  0.818997655  0.140777488
#> [37,] -0.259574965 1.1609341 -0.57057154 1.1958533  0.001143088 -0.686989444
#> [38,]  0.156069256 1.7401648  0.53605904 1.3051938  0.449312867 -0.333357949
#> [39,]  0.064540116 1.7615691  0.39733837 1.2702326  0.476533280 -0.238005935
#> [40,]  0.403904611 2.0898543  0.93044794 1.4114527  0.783964725  0.042013885
#>             [,7]        [,8]        [,9]       [,10]
#>  [1,] 0.34574843 -0.23004363 0.381158834  0.35223356
#>  [2,] 1.02407686 -0.29136987 0.612701895  0.95923551
#>  [3,] 0.93417216  0.34714235 0.865093753  0.93271412
#>  [4,] 1.07999119  0.15312835 0.735205695  0.99044550
#>  [5,] 0.50243587  0.08111916 0.522862361  0.56889064
#>  [6,] 1.16376442 -0.39170310 0.469164680  1.00739599
#>  [7,] 1.16758366  0.43262177 1.060381988  1.19441714
#>  [8,] 0.32005353 -0.16699674 0.265258094  0.27494834
#>  [9,] 0.93438083 -0.11414537 0.645425768  1.01474931
#> [10,] 1.13053159  0.12839684 0.662398129  1.05688941
#> [11,] 0.47478394 -0.25690930 0.413164045  0.36630223
#> [12,] 1.33089920  0.11228746 0.707026044  1.20789527
#> [13,] 0.65690309 -0.67697905 0.269060964  0.61663062
#> [14,] 0.80184436 -0.09976253 0.652560417  0.80208854
#> [15,] 0.49486713 -0.26060354 0.709813483  0.71680551
#> [16,] 1.03658792 -0.16868845 0.803415177  1.13640167
#> [17,] 1.17019823  0.12104649 0.869693320  1.14600738
#> [18,] 0.05792111 -0.70268822 0.009479391 -0.04543527
#> [19,] 0.75239028  0.17872198 0.840678169  0.92663914
#> [20,] 1.28352299  0.02722543 0.657760450  1.10257448
#> [21,] 1.22337992 -0.11067867 0.646839904  1.15971217
#> [22,] 0.65514145 -0.91697473 0.268980817  0.63834563
#> [23,] 0.46996075 -0.30230065 0.433440770  0.46984116
#> [24,] 1.09275910 -0.10858931 0.640456283  1.00959827
#> [25,] 1.09467478 -0.12388938 0.602891848  1.02057335
#> [26,] 0.38438426 -0.05007750 0.339773464  0.32880455
#> [27,] 1.57629820  0.09484279 0.944972819  1.47591121
#> [28,] 1.13833148 -0.24049470 0.480794529  0.98210902
#> [29,] 1.08285309  0.21167608 0.894173782  1.03834094
#> [30,] 1.57354242  0.10758038 0.982880645  1.50124187
#> [31,] 1.17020384 -0.47324399 0.445909781  1.02174293
#> [32,] 1.13372775  0.27527505 0.850275041  1.18763318
#> [33,] 0.79297156  0.27744610 0.796115520  0.80195110
#> [34,] 0.57598432  0.01237831 0.504551494  0.46642536
#> [35,] 0.37431406 -0.37863159 0.325081653  0.32614349
#> [36,] 1.06917247  0.25489449 0.959294394  1.17730913
#> [37,] 0.39790100 -0.62212346 0.189305728  0.37817839
#> [38,] 0.68628127 -0.35351555 0.606230982  0.82745408
#> [39,] 1.18455464 -0.29542776 0.475120618  0.99964693
#> [40,] 1.24057849 -0.02267232 0.877852615  1.29646991
#> 
#> $pre[[3]]
#>             [,1]          [,2]
#>  [1,] 0.08698471  0.0042980768
#>  [2,] 0.09148494 -0.0063926154
#>  [3,] 0.08406034 -0.0066761422
#>  [4,] 0.08715340 -0.0033276318
#>  [5,] 0.08625666  0.0009419743
#>  [6,] 0.09472667 -0.0027664342
#>  [7,] 0.08677602 -0.0072461779
#>  [8,] 0.08563046  0.0077733404
#>  [9,] 0.09118724 -0.0047928546
#> [10,] 0.08853543 -0.0020641369
#> [11,] 0.08547822  0.0010000800
#> [12,] 0.09175232 -0.0042154432
#> [13,] 0.09302743 -0.0051787632
#> [14,] 0.08659781 -0.0055816606
#> [15,] 0.09178032 -0.0060817963
#> [16,] 0.09611227 -0.0067668868
#> [17,] 0.08718424 -0.0100262396
#> [18,] 0.08304948  0.0010143884
#> [19,] 0.08634449 -0.0068917002
#> [20,] 0.08884417 -0.0050628298
#> [21,] 0.09028614 -0.0077549760
#> [22,] 0.09349359 -0.0097174915
#> [23,] 0.08402281 -0.0076498239
#> [24,] 0.09239274 -0.0044925713
#> [25,] 0.08810366 -0.0069471356
#> [26,] 0.08563107  0.0075624488
#> [27,] 0.09029866 -0.0111561667
#> [28,] 0.09206405 -0.0020336371
#> [29,] 0.08881599 -0.0031561091
#> [30,] 0.09023412 -0.0119778831
#> [31,] 0.09436459 -0.0034044473
#> [32,] 0.08827215 -0.0045958947
#> [33,] 0.08761443  0.0005493205
#> [34,] 0.08844609  0.0073904700
#> [35,] 0.08606159  0.0007578578
#> [36,] 0.08685230 -0.0095075709
#> [37,] 0.09512516  0.0081579357
#> [38,] 0.08723357 -0.0091168581
#> [39,] 0.09287328 -0.0023760523
#> [40,] 0.09016175 -0.0112779496
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4793401 -0.4793401
#>  [2,] -0.5244499  0.5244499
#>  [3,] -0.5226686  0.5226686
#>  [4,] -0.5226048  0.5226048
#>  [5,] -0.5213157  0.5213157
#>  [6,]  0.4756460 -0.4756460
#>  [7,]  0.4765118 -0.4765118
#>  [8,] -0.5194545  0.5194545
#>  [9,] -0.5239766  0.5239766
#> [10,] -0.5226344  0.5226344
#> [11,]  0.4788930 -0.4788930
#> [12,] -0.5239735  0.5239735
#> [13,]  0.4754682 -0.4754682
#> [14,] -0.5230286  0.5230286
#> [15,] -0.5244460  0.5244460
#> [16,] -0.5256971  0.5256971
#> [17,] -0.5242835  0.5242835
#> [18,]  0.4795027 -0.4795027
#> [19,]  0.4767078 -0.4767078
#> [20,]  0.4765405 -0.4765405
#> [21,] -0.5244907  0.5244907
#> [22,]  0.4742201 -0.4742201
#> [23,]  0.4770979 -0.4770979
#> [24,]  0.4757976 -0.4757976
#> [25,] -0.5237448  0.5237448
#> [26,] -0.5195072  0.5195072
#> [27,] -0.5253420  0.5253420
#> [28,]  0.4764929 -0.4764929
#> [29,] -0.5229768  0.5229768
#> [30,] -0.5255308  0.5255308
#> [31,] -0.5244228  0.5244228
#> [32,] -0.5232003  0.5232003
#> [33,] -0.5217525  0.5217525
#> [34,] -0.5202528  0.5202528
#> [35,]  0.4786870 -0.4786870
#> [36,]  0.4759287 -0.4759287
#> [37,] -0.5217281  0.5217281
#> [38,]  0.4759310 -0.4759310
#> [39,]  0.4762057 -0.4762057
#> [40,]  0.4746618 -0.4746618
#> 
#> $L
#> [1] 0.6905249 0.6998413 0.7493439 0.7068451 0.6861415 0.7009904
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174