Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13         V14
#>  [1,] -0.2668427 -0.09328489 -0.185002566 -0.187706585 -0.17404854 -0.09257008
#>  [2,] -0.2600643 -0.15426433 -0.202205723 -0.044154652 -0.14710815 -0.05576998
#>  [3,] -0.3903990 -0.16887184  0.008315282 -0.167170831 -0.08752674  0.02328697
#>  [4,] -0.2402565 -0.25354139 -0.149376052 -0.237188115 -0.10514297 -0.18979770
#>  [5,] -0.2232010 -0.18970870 -0.166376122 -0.131376229 -0.06703948 -0.05897942
#>  [6,] -0.1671419 -0.08859764 -0.121712496 -0.118210397 -0.05427972 -0.12751106
#>  [7,] -0.4030257 -0.09733102 -0.078733086 -0.071200230 -0.05969847 -0.10396083
#>  [8,] -0.3088466 -0.05601044 -0.018503425 -0.071144515 -0.10695418 -0.03168563
#>  [9,] -0.3650779 -0.13093167 -0.110689757 -0.001927387 -0.13385035 -0.04477493
#> [10,] -0.3670804 -0.06950314 -0.135482269 -0.133014840 -0.03956383 -0.05672435
#>               V15         V16         V17          V18         V19         V2
#>  [1,] -0.09924684  0.09383145  0.04009818 -0.058546498 -0.01584075 -0.2955969
#>  [2,] -0.13178884 -0.13490120  0.02087982 -0.052322789 -0.06718635 -0.2174771
#>  [3,] -0.13162106  0.02603267 -0.08520090  0.037770820  0.06849594 -0.2961209
#>  [4,] -0.06477090 -0.14095005 -0.16060226 -0.035567401 -0.07522608 -0.1648895
#>  [5,] -0.09183260 -0.13395702 -0.12186459 -0.027824812 -0.12842118 -0.3558658
#>  [6,] -0.06695167 -0.15690892  0.00140263 -0.007615327  0.04460672 -0.2443298
#>  [7,]  0.04338524 -0.03815650  0.10031807  0.003048916 -0.04048273 -0.3009548
#>  [8,] -0.11310920  0.04339947 -0.04327062 -0.048031169  0.02910517 -0.2350453
#>  [9,]  0.03006274  0.00563919  0.03846801  0.197764937  0.17345918 -0.3189772
#> [10,] -0.06843400  0.10036703  0.07923331  0.035061717  0.15472246 -0.3564193
#>                V20          V21         V22          V23           V24
#>  [1,]  0.138465092  0.019683732  0.05854917  0.077558698  0.1539869949
#>  [2,] -0.062537575  0.031338002  0.10400680 -0.005875158  0.0570833282
#>  [3,]  0.020929810  0.160661265  0.21835411  0.215029037  0.2575426694
#>  [4,]  0.007453415 -0.037829832 -0.05084646 -0.033549332  0.0001431552
#>  [5,] -0.081878283 -0.007179877  0.13781131  0.136894822  0.1282794925
#>  [6,]  0.076680507  0.079763657 -0.03395027  0.076652222  0.0674076624
#>  [7,]  0.059550042  0.074382813  0.20232547  0.141857563  0.2542038575
#>  [8,]  0.030089374  0.015239733  0.13365286  0.130302992 -0.0025012657
#>  [9,]  0.127886775  0.247763775  0.21434792  0.160728118  0.2997432859
#> [10,]  0.147223624  0.256773817  0.18573960  0.268479459  0.1850003472
#>               V25        V26       V27        V28          V29         V3
#>  [1,] 0.174736917 0.19763297 0.2139598 0.25992022  0.177507660 -0.2140964
#>  [2,] 0.031204306 0.20028754 0.1234477 0.09513463  0.127992605 -0.2555663
#>  [3,] 0.199976335 0.12301885 0.1805459 0.03670735  0.118183798 -0.2270645
#>  [4,] 0.102208518 0.16554418 0.1358115 0.15996559 -0.045805953 -0.1754610
#>  [5,] 0.145781796 0.07135943 0.0405723 0.14254537  0.036976175 -0.2813861
#>  [6,] 0.007082398 0.05694694 0.2030326 0.18268037 -0.016651244 -0.2208095
#>  [7,] 0.302942091 0.30650672 0.2235043 0.20341795  0.150981596 -0.3949449
#>  [8,] 0.142704306 0.03492401 0.1005051 0.02889731  0.135698409 -0.2985381
#>  [9,] 0.161376431 0.19120742 0.2145914 0.11906360  0.200297616 -0.2816081
#> [10,] 0.279188151 0.13512601 0.2086226 0.06154837 -0.007841149 -0.2862744
#>               V30           V31          V32         V33          V34
#>  [1,]  0.14623651 -1.859498e-02 -0.130330245 -0.11892441 -0.161201683
#>  [2,]  0.07103365  5.396006e-02  0.090568303 -0.01723381 -0.074341950
#>  [3,] -0.03121547  5.459640e-02  0.026982147  0.08240445  0.074340380
#>  [4,]  0.05963912 -2.528552e-03  0.074781645 -0.05372933  0.072138504
#>  [5,]  0.06525226 -1.482025e-02 -0.008321568 -0.05198448  0.034503464
#>  [6,] -0.08053507 -9.321385e-03 -0.102310274 -0.16471916 -0.035617702
#>  [7,]  0.20451854  6.626944e-05  0.024285266  0.10061889  0.003487802
#>  [8,] -0.06727101 -2.238316e-02 -0.088171040 -0.02083628  0.007605143
#>  [9,]  0.16865845  1.897607e-01  0.135128438  0.04193683  0.103026238
#> [10,]  0.03292468 -1.723438e-02 -0.069207858 -0.04336868 -0.022151200
#>                V35          V36          V37          V38         V39
#>  [1,] -0.131047675  0.030219911  0.075973166 -0.143131511  0.03214576
#>  [2,] -0.084826336  0.001408377  0.009269567 -0.107516080 -0.08129712
#>  [3,]  0.006939596  0.017136486 -0.007986575  0.084724316 -0.07994150
#>  [4,]  0.041023083 -0.068827869 -0.001592404 -0.028108421 -0.11536554
#>  [5,]  0.073918549 -0.080180732 -0.083446647 -0.048576682 -0.01648486
#>  [6,] -0.120323477 -0.101056093 -0.152367712 -0.120695454  0.01172273
#>  [7,]  0.054199776 -0.043471464 -0.016121292  0.046164191 -0.06940678
#>  [8,] -0.160791595  0.033257850 -0.042542767 -0.078970890 -0.08822057
#>  [9,]  0.093649027  0.143054366 -0.020275835  0.005887183 -0.04816260
#> [10,] -0.036826053  0.064709952  0.085458328 -0.087956271 -0.06279091
#>               V4          V40         V41         V42         V43         V44
#>  [1,] -0.2401044 -0.082786671  0.01660779 -0.20920160 -0.08911122 -0.10020013
#>  [2,] -0.2719986 -0.168269647 -0.03587477 -0.15018064 -0.18039502 -0.21939844
#>  [3,] -0.3243734 -0.055648299 -0.12927851 -0.02702476  0.01164320 -0.11273375
#>  [4,] -0.1901076 -0.051142756 -0.17905541 -0.03533332 -0.16905845 -0.23235465
#>  [5,] -0.3810716 -0.041171680 -0.06155801 -0.16514496 -0.03393750 -0.22933641
#>  [6,] -0.1718464 -0.056086272 -0.09724832 -0.19202376 -0.08840141 -0.08028872
#>  [7,] -0.1933617 -0.107575682 -0.08970601 -0.10983853 -0.08817010 -0.06117304
#>  [8,] -0.2932231 -0.006089574 -0.15777528 -0.23021393 -0.20628741 -0.15118583
#>  [9,] -0.3406705  0.067593122  0.04638278  0.08160422 -0.04196207 -0.12074279
#> [10,] -0.1657311 -0.132366660 -0.08131161 -0.01440586 -0.11176703 -0.12334520
#>               V45        V46         V47        V48        V49         V5
#>  [1,] -0.09036218 -0.1866749 -0.28739504 -0.1711642 -0.3228340 -0.2706721
#>  [2,] -0.22291864 -0.1611916 -0.20642574 -0.3076380 -0.2982111 -0.3277985
#>  [3,] -0.19996980 -0.1852606 -0.12121318 -0.1752347 -0.2821399 -0.3212553
#>  [4,] -0.21081841 -0.1083182 -0.21299297 -0.2897894 -0.1904583 -0.1725980
#>  [5,] -0.13359456 -0.2798461 -0.20361077 -0.2237834 -0.2476281 -0.3115721
#>  [6,] -0.07420364 -0.1287414 -0.15390984 -0.2766640 -0.2937732 -0.2787305
#>  [7,] -0.06465636 -0.1576121 -0.10336966 -0.2827049 -0.3275906 -0.2521966
#>  [8,] -0.11995291 -0.1165231 -0.25283747 -0.2299611 -0.2607417 -0.1899019
#>  [9,] -0.15329462 -0.1594915 -0.06436173 -0.1320814 -0.1920567 -0.2327398
#> [10,] -0.19754207 -0.1219370 -0.18915969 -0.2900722 -0.2655495 -0.1576354
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2884271 -0.4087896 -0.2770317 -0.3229304 -0.3231451 -0.4302767
#>  [2,] -0.3633220 -0.3387371 -0.2820511 -0.3050139 -0.2744771 -0.4146559
#>  [3,] -0.4016625 -0.3406349 -0.4003100 -0.3511801 -0.3874822 -0.3124104
#>  [4,] -0.2012865 -0.3082024 -0.2482964 -0.2916868 -0.2954272 -0.3513843
#>  [5,] -0.2835789 -0.3786956 -0.3353870 -0.2780889 -0.3374610 -0.3943975
#>  [6,] -0.2932222 -0.3216875 -0.2079114 -0.2451770 -0.2062645 -0.3212978
#>  [7,] -0.2656437 -0.3927120 -0.4249188 -0.3275720 -0.4277002 -0.2684327
#>  [8,] -0.2425646 -0.2796059 -0.3228130 -0.3111465 -0.2962279 -0.2714555
#>  [9,] -0.4077268 -0.4025215 -0.3959270 -0.4119504 -0.3742061 -0.3874615
#> [10,] -0.2663602 -0.2708978 -0.3651689 -0.4039134 -0.2750925 -0.3530578
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4654533 -0.3106445 -0.3299710 -0.3552679 -0.2265337 -0.3694094
#>  [2,] -0.2787300 -0.4132341 -0.3265222 -0.3224858 -0.2133657 -0.3498320
#>  [3,] -0.3954843 -0.2765828 -0.4034657 -0.2601358 -0.2054315 -0.3655281
#>  [4,] -0.2298056 -0.2083464 -0.3078131 -0.2374622 -0.1289660 -0.3079631
#>  [5,] -0.2837057 -0.4008256 -0.2684408 -0.4194845 -0.2323143 -0.3654083
#>  [6,] -0.1873624 -0.1723956 -0.2166646 -0.2464435 -0.1514057 -0.3319274
#>  [7,] -0.3707681 -0.3759658 -0.3904577 -0.3491160 -0.1780531 -0.3518989
#>  [8,] -0.3224313 -0.2833096 -0.2731943 -0.3533629 -0.1555093 -0.2604008
#>  [9,] -0.2619594 -0.4121746 -0.4164217 -0.4353711 -0.1026434 -0.2969307
#> [10,] -0.3539451 -0.3582562 -0.3259317 -0.3606336 -0.2132453 -0.4221832
#>                V7          V8          V9
#>  [1,] -0.22181094 -0.07509987 -0.08013626
#>  [2,] -0.28244473 -0.21516755 -0.23543526
#>  [3,] -0.19023464 -0.06711611 -0.05110991
#>  [4,] -0.19844179 -0.14256139 -0.18016842
#>  [5,] -0.14819639 -0.21147760 -0.29576821
#>  [6,] -0.13097862 -0.23352086 -0.07784599
#>  [7,] -0.13526391 -0.13432412 -0.14662628
#>  [8,] -0.16785835 -0.25689712 -0.08655177
#>  [9,] -0.08787346 -0.12949867 -0.17836275
#> [10,] -0.21358427 -0.22774664 -0.06292659
#> 
#> $W[[2]]
#>             [,1]         [,2]        [,3]         [,4]        [,5]        [,6]
#> [1,] 0.143048013 -0.008769375 0.173052106 -0.004137596  0.04859173  0.06160981
#> [2,] 0.009263136 -0.084033307 0.001837344  0.028031091 -0.07334019 -0.05069841
#>             [,7]        [,8]       [,9]       [,10]
#> [1,]  0.15762120 -0.03853321  0.1369153  0.16494334
#> [2,] -0.04140473 -0.08982032 -0.1581015 -0.09232397
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.995713e-04  0.0014594701  0.0017303444  0.0017140423  0.0016337315
#>  [2,]  1.431086e-04  0.0010535772  0.0012718357  0.0013267206  0.0013383102
#>  [3,]  1.707909e-04  0.0012025668  0.0014059115  0.0013861928  0.0013250273
#>  [4,]  2.532039e-04  0.0020540266  0.0025320830  0.0025969257  0.0024694391
#>  [5,] -2.046276e-05 -0.0003173375 -0.0004281781 -0.0003771040 -0.0002011930
#>  [6,]  4.226750e-05  0.0002817650  0.0003257989  0.0003486744  0.0003432805
#>  [7,]  1.512025e-04  0.0010190205  0.0011754786  0.0011289765  0.0010805809
#>  [8,]  2.869138e-04  0.0022741980  0.0027677094  0.0027853025  0.0026037799
#>  [9,]  2.558242e-05  0.0001069060  0.0001056638  0.0001268499  0.0001409768
#> [10,]  4.546365e-05  0.0002168004  0.0002292840  0.0002632487  0.0002760693
#>                 V14          V15          V16          V17          V18
#>  [1,]  1.525142e-03 1.491379e-03 0.0016485242 0.0016167010 0.0017880816
#>  [2,]  1.277852e-03 1.269718e-03 0.0014429824 0.0014909920 0.0016513189
#>  [3,]  1.247258e-03 1.238670e-03 0.0013673043 0.0013208429 0.0014552084
#>  [4,]  2.214188e-03 2.079677e-03 0.0023090316 0.0023599531 0.0026662954
#>  [5,] -5.188618e-05 7.695794e-05 0.0001542503 0.0002024319 0.0001685229
#>  [6,]  3.235366e-04 3.385944e-04 0.0003967948 0.0004092773 0.0004628293
#>  [7,]  1.040986e-03 1.063694e-03 0.0011714255 0.0011032430 0.0011943564
#>  [8,]  2.335118e-03 2.196403e-03 0.0024140091 0.0024140342 0.0027404513
#>  [9,]  1.257646e-04 1.773267e-04 0.0002231384 0.0001954662 0.0002137498
#> [10,]  2.512760e-04 3.141296e-04 0.0003874794 0.0003602086 0.0003881228
#>                 V19            V2           V20          V21          V22
#>  [1,]  2.335913e-03  2.575973e-04  2.895320e-03 3.131934e-03 0.0028774545
#>  [2,]  2.022124e-03  1.870464e-04  2.371289e-03 2.599271e-03 0.0025207635
#>  [3,]  1.898934e-03  2.242518e-04  2.389659e-03 2.596872e-03 0.0023852278
#>  [4,]  3.424051e-03  3.064846e-04  4.030266e-03 4.351733e-03 0.0039704766
#>  [5,] -6.155959e-06 -1.302476e-06 -6.409636e-05 2.501082e-05 0.0002822741
#>  [6,]  5.811047e-04  6.786072e-05  7.231951e-04 8.167567e-04 0.0008775779
#>  [7,]  1.565297e-03  1.992854e-04  2.012894e-03 2.177383e-03 0.0019510467
#>  [8,]  3.622659e-03  3.520830e-04  4.374240e-03 4.689160e-03 0.0042395624
#>  [9,]  2.614191e-04  5.174849e-05  3.299717e-04 3.890444e-04 0.0004642529
#> [10,]  4.661427e-04  8.280542e-05  5.793753e-04 6.826958e-04 0.0007801135
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0027881693 0.0027907849 0.0026782705 0.0029324281 0.0030402892
#>  [2,] 0.0025157739 0.0025259071 0.0024836340 0.0026841374 0.0027637064
#>  [3,] 0.0023304966 0.0023450664 0.0022770852 0.0025181288 0.0026236895
#>  [4,] 0.0037863644 0.0038647438 0.0037512701 0.0040461241 0.0041812139
#>  [5,] 0.0004629892 0.0004457867 0.0005515179 0.0006072749 0.0006133252
#>  [6,] 0.0009333504 0.0009304838 0.0009289711 0.0010150362 0.0010762682
#>  [7,] 0.0018737419 0.0018709424 0.0017971665 0.0019966194 0.0020676042
#>  [8,] 0.0040287740 0.0040622767 0.0038697455 0.0041936536 0.0043392003
#>  [9,] 0.0005507258 0.0005644924 0.0005757735 0.0006287198 0.0006586830
#> [10,] 0.0008836924 0.0009063684 0.0009257259 0.0010241168 0.0010855021
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0031334316 0.0027088941  2.661908e-04 0.0023669984 0.0018017493
#>  [2,] 0.0027763607 0.0024259522  1.965902e-04 0.0021093327 0.0017114224
#>  [3,] 0.0027281803 0.0023863055  2.296756e-04 0.0021129730 0.0016225577
#>  [4,] 0.0042426991 0.0035481669  3.221406e-04 0.0029316766 0.0022270473
#>  [5,] 0.0005519473 0.0006261098 -6.487708e-07 0.0007043211 0.0007662465
#>  [6,] 0.0010681346 0.0009617240  7.052513e-05 0.0008504409 0.0006707247
#>  [7,] 0.0021724103 0.0019042270  2.025267e-04 0.0017249203 0.0013055954
#>  [8,] 0.0044544098 0.0037563997  3.664451e-04 0.0031102338 0.0022660969
#>  [9,] 0.0006885271 0.0006616568  5.216071e-05 0.0006300318 0.0005329051
#> [10,] 0.0011380516 0.0010994056  8.274540e-05 0.0010515039 0.0008881682
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0016460440 0.0014385967 0.0014112652 0.0012451372 0.0009430233
#>  [2,] 0.0015201611 0.0013268673 0.0012948108 0.0011897803 0.0010091595
#>  [3,] 0.0014812767 0.0012951808 0.0012814117 0.0011778873 0.0009417944
#>  [4,] 0.0019318536 0.0015640242 0.0014064348 0.0010764179 0.0005609380
#>  [5,] 0.0006916570 0.0006997525 0.0007736108 0.0009415584 0.0011700248
#>  [6,] 0.0006161079 0.0005404595 0.0005343780 0.0005393199 0.0005523011
#>  [7,] 0.0012123271 0.0010922860 0.0011105554 0.0010409591 0.0008371525
#>  [8,] 0.0020179499 0.0016739757 0.0015588157 0.0012043633 0.0006523806
#>  [9,] 0.0005157208 0.0004685493 0.0004454110 0.0004801018 0.0005694618
#> [10,] 0.0008351940 0.0007462855 0.0007219070 0.0007768854 0.0008682870
#>                V37          V38          V39            V4          V40
#>  [1,] 0.0009849364 0.0013365014 0.0015233230  3.615076e-04 0.0012822334
#>  [2,] 0.0010057576 0.0011902386 0.0013015385  2.656476e-04 0.0011745780
#>  [3,] 0.0009609151 0.0012270789 0.0013367711  3.102261e-04 0.0011303417
#>  [4,] 0.0007272843 0.0013115517 0.0017529551  4.627249e-04 0.0015264851
#>  [5,] 0.0010011060 0.0006969988 0.0004320801 -2.470894e-05 0.0004868941
#>  [6,] 0.0005473576 0.0005454840 0.0004986174  8.393849e-05 0.0004473537
#>  [7,] 0.0008399355 0.0010712540 0.0011505480  2.737589e-04 0.0009513120
#>  [8,] 0.0008358842 0.0014999909 0.0019583627  5.176362e-04 0.0016311996
#>  [9,] 0.0005735193 0.0005182563 0.0004490780  5.397184e-05 0.0004363331
#> [10,] 0.0008545198 0.0007972423 0.0007057777  8.570927e-05 0.0006629460
#>                V41          V42          V43          V44           V45
#>  [1,] 0.0011674105 0.0013730566 0.0013506526 1.234331e-03  1.323957e-03
#>  [2,] 0.0010942770 0.0011988877 0.0011246741 9.871193e-04  9.892639e-04
#>  [3,] 0.0010402676 0.0011834251 0.0011655305 1.060686e-03  1.150231e-03
#>  [4,] 0.0013407652 0.0017811075 0.0017416156 1.636128e-03  1.643431e-03
#>  [5,] 0.0005530224 0.0002842785 0.0001531862 1.496408e-05 -6.513097e-06
#>  [6,] 0.0004487515 0.0004185949 0.0004036659 3.377917e-04  3.527842e-04
#>  [7,] 0.0008647335 0.0009809324 0.0009632781 8.691923e-04  9.760965e-04
#>  [8,] 0.0014014347 0.0018563841 0.0018835823 1.781042e-03  1.836511e-03
#>  [9,] 0.0004469879 0.0003614336 0.0003054468 2.302628e-04  2.690844e-04
#> [10,] 0.0006786993 0.0005774829 0.0005165337 4.124688e-04  4.581330e-04
#>                V46          V47           V48           V49            V5
#>  [1,] 1.050964e-03 7.260842e-04  5.784455e-04  3.287867e-04  4.156481e-04
#>  [2,] 7.886880e-04 5.798851e-04  4.493194e-04  2.561628e-04  3.355043e-04
#>  [3,] 9.227031e-04 6.284511e-04  4.911981e-04  2.778453e-04  3.486147e-04
#>  [4,] 1.297700e-03 9.782242e-04  8.026085e-04  4.475407e-04  5.939108e-04
#>  [5,] 1.812176e-05 1.183407e-05 -4.035196e-05 -1.686110e-05 -3.020579e-05
#>  [6,] 2.934104e-04 1.981935e-04  1.322317e-04  7.880114e-05  9.042134e-05
#>  [7,] 7.846580e-04 5.208111e-04  4.103253e-04  2.326704e-04  2.945149e-04
#>  [8,] 1.441271e-03 1.043188e-03  8.585150e-04  4.851733e-04  6.331179e-04
#>  [9,] 2.424013e-04 1.564577e-04  8.794933e-05  5.784400e-05  5.425638e-05
#> [10,] 3.945966e-04 2.546344e-04  1.520838e-04  9.526850e-05  8.863208e-05
#>                V50           V51          V52          V53          V54
#>  [1,] 9.486733e-05  9.851103e-05 8.710655e-05 5.401410e-05 6.482340e-05
#>  [2,] 7.664222e-05  7.686702e-05 6.567344e-05 4.350676e-05 4.938006e-05
#>  [3,] 8.517770e-05  8.449456e-05 7.651542e-05 4.784408e-05 5.548176e-05
#>  [4,] 1.002463e-04  1.303256e-04 1.036692e-04 6.412668e-05 7.691804e-05
#>  [5,] 2.569094e-05 -1.553532e-06 4.048346e-06 7.508425e-06 3.617231e-06
#>  [6,] 3.367273e-05  2.444927e-05 2.479960e-05 1.616713e-05 1.749369e-05
#>  [7,] 7.334175e-05  7.146246e-05 6.602971e-05 4.078804e-05 4.850720e-05
#>  [8,] 1.176146e-04  1.419028e-04 1.177539e-04 7.275764e-05 8.790387e-05
#>  [9,] 3.100585e-05  1.452529e-05 1.693412e-05 1.140296e-05 1.046917e-05
#> [10,] 4.738604e-05  2.487473e-05 2.755690e-05 1.887387e-05 1.844384e-05
#>                V55          V56          V57          V58          V59
#>  [1,] 4.947941e-05 4.326458e-05 2.688667e-05 4.061281e-05 4.213953e-05
#>  [2,] 3.920385e-05 3.460561e-05 2.270246e-05 3.238003e-05 3.498974e-05
#>  [3,] 4.195832e-05 3.691476e-05 2.394661e-05 3.565002e-05 3.650591e-05
#>  [4,] 6.419306e-05 5.786250e-05 3.202418e-05 4.570474e-05 5.220010e-05
#>  [5,] 7.843150e-07 7.615980e-07 6.279406e-06 7.382694e-06 6.826743e-06
#>  [6,] 1.027002e-05 1.043507e-05 7.603243e-06 1.177602e-05 1.155117e-05
#>  [7,] 3.675108e-05 3.165023e-05 2.101971e-05 3.175806e-05 3.173156e-05
#>  [8,] 7.025478e-05 6.233771e-05 3.453066e-05 5.232601e-05 5.644545e-05
#>  [9,] 7.248939e-06 6.362735e-06 5.054677e-06 1.112627e-05 9.123999e-06
#> [10,] 1.244310e-05 1.114779e-05 8.626943e-06 1.653427e-05 1.501570e-05
#>                 V6          V60           V7            V8            V9
#>  [1,] 5.113471e-04 3.083636e-05 5.597837e-04  7.987512e-04  0.0012518215
#>  [2,] 4.148918e-04 2.516741e-05 4.531015e-04  5.854165e-04  0.0008742368
#>  [3,] 4.367280e-04 2.769916e-05 4.868433e-04  6.750681e-04  0.0010444264
#>  [4,] 6.706259e-04 3.515073e-05 7.019246e-04  1.044982e-03  0.0016819623
#>  [5,] 2.011955e-05 6.809942e-06 5.200613e-05 -8.204789e-05 -0.0002495972
#>  [6,] 1.413030e-04 9.869854e-06 1.655633e-04  1.772364e-04  0.0002532949
#>  [7,] 3.636149e-04 2.423381e-05 4.111174e-04  5.831412e-04  0.0008972709
#>  [8,] 7.337706e-04 3.949388e-05 7.835449e-04  1.171118e-03  0.0018872111
#>  [9,] 9.849205e-05 8.795087e-06 1.256234e-04  1.137296e-04  0.0001124096
#> [10,] 1.542525e-04 1.360146e-05 1.965944e-04  1.917583e-04  0.0002176737
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,] -0.09227539 -0.05963337 -0.1062573 -0.03994539 -0.04921834 -0.05449863
#> [2,]  0.09227539  0.05963337  0.1062573  0.03994539  0.04921834  0.05449863
#>            [,7]       [,8]       [,9]      [,10]
#> [1,] -0.1185442 -0.0724017 -0.1306495 -0.1165518
#> [2,]  0.1185442  0.0724017  0.1306495  0.1165518
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.47080536  0.46728702  0.45196043  0.15519771  0.21612131 -0.02425109
#>  [7]  0.55964393  0.38577542  0.63161324  0.68414950
#> 
#> $B[[2]]
#> [1]  0.2373092 -0.2512316
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.004177304 0.003722619 0.003627953 0.005306338 0.001053122 0.001417874
#>  [7] 0.002999827 0.005662472 0.001015733 0.001612422
#> 
#> $vB[[2]]
#> [1] -0.1501355  0.1501355
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#>  [2,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#>  [3,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>  [4,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#>  [5,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#>  [6,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#>  [7,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#>  [8,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [9,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [10,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [11,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [12,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [13,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [14,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [15,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [16,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [17,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [18,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [19,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [20,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [21,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [22,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [23,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [24,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [25,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [26,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [27,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [28,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [29,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [30,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [31,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [32,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [33,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [34,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [35,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [36,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [37,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [38,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [39,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [40,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#>  [2,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#>  [3,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>  [4,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#>  [5,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#>  [6,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#>  [7,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#>  [8,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [9,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [10,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [11,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [12,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [13,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [14,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [15,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [16,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [17,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [18,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [19,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [20,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [21,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [22,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [23,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [24,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [25,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [26,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [27,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [28,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [29,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [30,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [31,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [32,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [33,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [34,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [35,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [36,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [37,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [38,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [39,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [40,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#>  [2,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#>  [3,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>  [4,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#>  [5,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#>  [6,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#>  [7,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#>  [8,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [9,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [10,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [11,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [12,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [13,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [14,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [15,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [16,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [17,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [18,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [19,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [20,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [21,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [22,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [23,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [24,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [25,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [26,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [27,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [28,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [29,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [30,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [31,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [32,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [33,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [34,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [35,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [36,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [37,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [38,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [39,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [40,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#>  [2,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#>  [3,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>  [4,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#>  [5,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#>  [6,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#>  [7,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#>  [8,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [9,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [10,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [11,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [12,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [13,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [14,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [15,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [16,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [17,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [18,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [19,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [20,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [21,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [22,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [23,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [24,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [25,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [26,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [27,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [28,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [29,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [30,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [31,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [32,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [33,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [34,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [35,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [36,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [37,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [38,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [39,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [40,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#>  [2,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#>  [3,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>  [4,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#>  [5,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#>  [6,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#>  [7,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#>  [8,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [9,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [10,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [11,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [12,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [13,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [14,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [15,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [16,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [17,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [18,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [19,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [20,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [21,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [22,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [23,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [24,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [25,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [26,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [27,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [28,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [29,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [30,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [31,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [32,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [33,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [34,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [35,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [36,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [37,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [38,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [39,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [40,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#>  [2,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#>  [3,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#>  [4,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#>  [5,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#>  [6,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#>  [7,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#>  [8,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [9,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [10,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [11,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [12,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [13,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [14,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [15,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [16,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [17,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [18,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [19,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [20,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [21,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [22,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [23,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [24,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [25,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [26,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [27,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [28,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [29,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [30,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [31,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [32,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [33,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [34,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [35,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [36,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [37,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [38,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [39,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [40,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6909426 0.5136889 0.7469806 0.4288382 0.4732942 0.3967379 0.8116411
#>  [2,] 0.6997180 0.5691068 0.8027610 0.4214413 0.5196748 0.4483802 0.8251283
#>  [3,] 0.6702544 0.4804766 0.7497774 0.3448395 0.4317491 0.3951197 0.8294697
#>  [4,] 0.7481909 0.6142963 0.7898541 0.5125331 0.5777205 0.4689806 0.8534421
#>  [5,] 0.7332076 0.6097346 0.7588582 0.4932660 0.5309694 0.4468880 0.8500976
#>  [6,] 0.6584070 0.4324804 0.7232016 0.3092070 0.3658999 0.3106245 0.7963172
#>  [7,] 0.7137911 0.6057170 0.7446311 0.4941724 0.5322207 0.4259096 0.8310340
#>  [8,] 0.4846250 0.3841342 0.6202369 0.3010397 0.3173527 0.2260933 0.6784153
#>  [9,] 0.6942003 0.5992658 0.8180027 0.4977726 0.5825188 0.3824692 0.8685218
#> [10,] 0.6446445 0.5127529 0.7148590 0.4234528 0.4382949 0.3093946 0.8263163
#> [11,] 0.4030842 0.3106454 0.5146181 0.2664933 0.2878963 0.1811364 0.5853975
#> [12,] 0.5752614 0.3998193 0.6937149 0.2751440 0.3437859 0.3354081 0.7593093
#> [13,] 0.5012948 0.4026528 0.6251217 0.2965872 0.3315340 0.3264897 0.6681071
#> [14,] 0.7212404 0.5741210 0.7645263 0.4682985 0.5322747 0.4666165 0.8208461
#> [15,] 0.6751528 0.5020458 0.7887866 0.3766893 0.4591076 0.3985139 0.8274420
#> [16,] 0.7897085 0.6669959 0.8492397 0.5421019 0.6087843 0.4698205 0.9042401
#> [17,] 0.6496429 0.4224643 0.7302386 0.3054963 0.3772336 0.3706991 0.8081153
#> [18,] 0.6428764 0.4270780 0.7201653 0.2780292 0.3536814 0.3659559 0.7932790
#> [19,] 0.7785350 0.6519181 0.8165145 0.5175731 0.5704800 0.4804665 0.8892906
#> [20,] 0.7933363 0.6802143 0.8523273 0.5635692 0.6190492 0.4767581 0.9086851
#> [21,] 0.6629002 0.4917738 0.7648469 0.3499688 0.4288720 0.3932763 0.8165853
#> [22,] 0.7692845 0.6217829 0.8271006 0.4842748 0.5640998 0.4534394 0.8906847
#> [23,] 0.7029043 0.5848213 0.7898537 0.4612919 0.5115411 0.3274457 0.8743985
#> [24,] 0.7310414 0.6181958 0.7632741 0.5056580 0.5371155 0.4418516 0.8522174
#> [25,] 0.7791473 0.5833332 0.8212923 0.4172079 0.5129682 0.4604269 0.8886482
#> [26,] 0.5715601 0.4068701 0.7410723 0.3033095 0.4081172 0.3063340 0.7723008
#> [27,] 0.7816403 0.6050015 0.8219511 0.4524947 0.5285490 0.4587380 0.8941408
#> [28,] 0.4794715 0.3133646 0.5960868 0.2843305 0.2735506 0.2218580 0.6392217
#> [29,] 0.7657981 0.6076413 0.8040207 0.4676414 0.5338951 0.4782420 0.8832861
#> [30,] 0.5381185 0.4373547 0.7160065 0.3137219 0.3936915 0.3440085 0.7364533
#> [31,] 0.7136067 0.5978701 0.7487544 0.4840411 0.5219116 0.4293439 0.8456653
#> [32,] 0.6610265 0.4089958 0.7530590 0.2770489 0.3735837 0.3642564 0.8091070
#> [33,] 0.5348711 0.4335987 0.7496155 0.3295769 0.4180553 0.2843190 0.7972603
#> [34,] 0.6399316 0.5368264 0.8196936 0.4223480 0.4992843 0.3574820 0.8545660
#> [35,] 0.7473778 0.5363269 0.8019853 0.4039701 0.4951800 0.4389461 0.8627266
#> [36,] 0.6330732 0.5127626 0.7847745 0.3891452 0.4894356 0.3775017 0.8121094
#> [37,] 0.4670061 0.2973884 0.5722160 0.2453252 0.2437684 0.2163515 0.6438557
#> [38,] 0.7897616 0.6845774 0.8424786 0.5733959 0.6215734 0.4825572 0.9012839
#> [39,] 0.4308726 0.3328060 0.5504691 0.2712767 0.2811498 0.2278805 0.6543179
#> [40,] 0.7601772 0.5832381 0.8086459 0.4466253 0.5193162 0.4773497 0.8807616
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5424129 0.8964869 0.8123995
#>  [2,] 0.6194005 0.9011301 0.8666350
#>  [3,] 0.5340989 0.8904193 0.7859370
#>  [4,] 0.6073443 0.8864334 0.8236859
#>  [5,] 0.5661781 0.8910103 0.7904827
#>  [6,] 0.4775937 0.8778240 0.8081300
#>  [7,] 0.5670671 0.8810136 0.7653779
#>  [8,] 0.3592035 0.8601408 0.6282398
#>  [9,] 0.5677074 0.9270260 0.8211163
#> [10,] 0.4478611 0.8876781 0.7017102
#> [11,] 0.2887546 0.7767646 0.4885181
#> [12,] 0.4732897 0.8662464 0.7850104
#> [13,] 0.4684949 0.8155358 0.7299267
#> [14,] 0.5968631 0.8684863 0.8134572
#> [15,] 0.5509200 0.9184141 0.8566562
#> [16,] 0.6430438 0.9373819 0.8776902
#> [17,] 0.4969299 0.8804183 0.7770620
#> [18,] 0.5293835 0.8899594 0.8232480
#> [19,] 0.6144026 0.9190764 0.8543852
#> [20,] 0.6453959 0.9396792 0.8767808
#> [21,] 0.5522096 0.9066434 0.8452933
#> [22,] 0.6120705 0.9252887 0.8607555
#> [23,] 0.5106263 0.9317469 0.8019860
#> [24,] 0.5661630 0.8934207 0.7869521
#> [25,] 0.6037562 0.9348935 0.8810283
#> [26,] 0.4824638 0.8665942 0.7653082
#> [27,] 0.6009881 0.9340110 0.8693088
#> [28,] 0.3503855 0.7995466 0.6268954
#> [29,] 0.6064111 0.9132160 0.8467160
#> [30,] 0.5124572 0.8723911 0.7980849
#> [31,] 0.5573832 0.8889063 0.7751910
#> [32,] 0.5230798 0.8934330 0.8342113
#> [33,] 0.4481759 0.9198805 0.7527597
#> [34,] 0.5371615 0.9386789 0.8426904
#> [35,] 0.5728052 0.9271516 0.8745118
#> [36,] 0.5321883 0.9183107 0.8117096
#> [37,] 0.3477360 0.8136098 0.6224005
#> [38,] 0.6402144 0.9335013 0.8675394
#> [39,] 0.3026559 0.7952014 0.5551759
#> [40,] 0.6069659 0.9193067 0.8564289
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5080889 0.4919111
#>  [2,] 0.5057534 0.4942466
#>  [3,] 0.5106087 0.4893913
#>  [4,] 0.5022098 0.4977902
#>  [5,] 0.5038492 0.4961508
#>  [6,] 0.5137654 0.4862346
#>  [7,] 0.5041760 0.4958240
#>  [8,] 0.5200863 0.4799137
#>  [9,] 0.5042495 0.4957505
#> [10,] 0.5102187 0.4897813
#> [11,] 0.5247535 0.4752465
#> [12,] 0.5165423 0.4834577
#> [13,] 0.5176964 0.4823036
#> [14,] 0.5045676 0.4954324
#> [15,] 0.5094674 0.4905326
#> [16,] 0.4992825 0.5007175
#> [17,] 0.5134046 0.4865954
#> [18,] 0.5139124 0.4860876
#> [19,] 0.5009158 0.4990842
#> [20,] 0.4984404 0.5015596
#> [21,] 0.5104546 0.4895454
#> [22,] 0.5022605 0.4977395
#> [23,] 0.5063507 0.4936493
#> [24,] 0.5034020 0.4965980
#> [25,] 0.5046196 0.4953804
#> [26,] 0.5151347 0.4848653
#> [27,] 0.5033787 0.4966213
#> [28,] 0.5214447 0.4785553
#> [29,] 0.5030238 0.4969762
#> [30,] 0.5150242 0.4849758
#> [31,] 0.5046611 0.4953389
#> [32,] 0.5137767 0.4862233
#> [33,] 0.5154595 0.4845405
#> [34,] 0.5082859 0.4917141
#> [35,] 0.5068532 0.4931468
#> [36,] 0.5100480 0.4899520
#> [37,] 0.5230444 0.4769556
#> [38,] 0.4983674 0.5016326
#> [39,] 0.5232327 0.4767673
#> [40,] 0.5039393 0.4960607
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]         [,2]      [,3]         [,4]         [,5]        [,6]
#>  [1,]  0.804529807  0.054769449 1.0825729 -0.286592769 -0.106925042 -0.41907594
#>  [2,]  0.845955337  0.278207677 1.4036407 -0.316859698  0.078740027 -0.20721769
#>  [3,]  0.709335941 -0.078133412 1.0974257 -0.641801092 -0.274718212 -0.42584168
#>  [4,]  1.088986897  0.465408113 1.3240463  0.050143063  0.313422956 -0.12423720
#>  [5,]  1.010958061  0.446196733 1.1464298 -0.026937575  0.124036383 -0.21325245
#>  [6,]  0.656203512 -0.271738462 0.9603985 -0.803829127 -0.549847031 -0.79720115
#>  [7,]  0.913868332  0.429344137 1.0701798 -0.023311650  0.129061603 -0.29855969
#>  [8,] -0.061519510 -0.472037295 0.4905537 -0.842351644 -0.765964730 -1.23050386
#>  [9,]  0.819830127  0.402406990 1.5028736 -0.008909608  0.333122133 -0.47908099
#> [10,]  0.595580284  0.051022472 0.9191013 -0.308615205 -0.248084936 -0.80295099
#> [11,] -0.392630568 -0.797103647 0.0584892 -1.012487412 -0.905623387 -1.50866713
#> [12,]  0.303350694 -0.406218219 0.8175447 -0.968678282 -0.646468177 -0.68382529
#> [13,]  0.005179377 -0.394423850 0.5113449 -0.863602856 -0.701255195 -0.72410489
#> [14,]  0.950622964  0.298684875 1.1776571 -0.126976319  0.129278407 -0.13373300
#> [15,]  0.731584211  0.008183398 1.3176266 -0.503624405 -0.163935583 -0.41166120
#> [16,]  1.323168947  0.694629212 1.7286502  0.168807375  0.442204972 -0.12086512
#> [17,]  0.617470049 -0.312665383 0.9958333 -0.821259813 -0.501306859 -0.52921865
#> [18,]  0.587870648 -0.293782998 0.9452818 -0.954258697 -0.602895851 -0.54960579
#> [19,]  1.257149282  0.627481242 1.4929089  0.070321542  0.283809794 -0.07817385
#> [20,]  1.345154181  0.754756637 1.7529726  0.255660294  0.485514708 -0.09303475
#> [21,]  0.676245351 -0.032907766 1.1794390 -0.619176445 -0.286454721 -0.43356113
#> [22,]  1.204275369  0.497122502 1.5652162 -0.062921665  0.257817897 -0.18678360
#> [23,]  0.861166405  0.342597114 1.3240436 -0.155142683  0.046172628 -0.71976070
#> [24,]  0.999912446  0.481897153 1.1707144  0.022632824  0.148735729 -0.23365071
#> [25,]  1.260704152  0.336471882 1.5251272 -0.334245989  0.051884252 -0.15862427
#> [26,]  0.288219309 -0.376919649 1.0515494 -0.831587625 -0.371754131 -0.81731461
#> [27,]  1.275251194  0.426348832 1.5296228 -0.190596064  0.114320256 -0.16542435
#> [28,] -0.082160238 -0.784435925 0.3891865 -0.923081222 -0.976682338 -1.25487135
#> [29,]  1.184734793  0.437408472 1.4116161 -0.129615536  0.135788717 -0.08708710
#> [30,]  0.152770250 -0.251905018 0.9247377 -0.782775799 -0.431821372 -0.64548169
#> [31,]  0.912965946  0.396598532 1.0919799 -0.063857421  0.087702466 -0.28452855
#> [32,]  0.667872055 -0.368118277 1.1149943 -0.959147651 -0.516873145 -0.55693727
#> [33,]  0.139711241 -0.267183224 1.0965627 -0.710099337 -0.330761718 -0.92313776
#> [34,]  0.575067086  0.147572870 1.5142730 -0.313141894 -0.002862676 -0.58630967
#> [35,]  1.084675930  0.145564196 1.3987493 -0.388950155 -0.019280742 -0.24544040
#> [36,]  0.545423831  0.051061407 1.2937103 -0.450906943 -0.042263883 -0.50016570
#> [37,] -0.132167642 -0.859765363 0.2908981 -1.123702138 -1.132129314 -1.28705632
#> [38,]  1.323488915  0.774888230 1.6767869  0.295720120  0.496231866 -0.06979935
#> [39,] -0.278291658 -0.695521250 0.2025662 -0.988154665 -0.938765191 -1.22031779
#> [40,]  1.153651105  0.336080490 1.4412353 -0.214315182  0.077303133 -0.09066344
#>            [,7]        [,8]     [,9]       [,10]
#>  [1,] 1.4607093  0.17006009 2.158786  1.46567758
#>  [2,] 1.5514860  0.48700451 2.209845  1.87152796
#>  [3,] 1.5818737  0.13660749 2.095032  1.30060604
#>  [4,] 1.7618570  0.43616266 2.054817  1.54152249
#>  [5,] 1.7353670  0.26627460 2.101102  1.32783738
#>  [6,] 1.3634339 -0.08968507 1.971983  1.43790494
#>  [7,] 1.5929731  0.26989502 2.002064  1.18239383
#>  [8,] 0.7464987 -0.57882296 1.816460  0.52467273
#>  [9,] 1.8879513  0.27250338 2.541878  1.52392899
#> [10,] 1.5597415 -0.20931651 2.067240  0.85545514
#> [11,] 0.3449710 -0.90144037 1.246910 -0.04593587
#> [12,] 1.1488963 -0.10694299 1.868170  1.29510754
#> [13,] 0.6996362 -0.12618755 1.486390  0.99425090
#> [14,] 1.5220906  0.39241173 1.887641  1.47263280
#> [15,] 1.5676057  0.20438865 2.420992  1.78779082
#> [16,] 2.2452515  0.58859978 2.706036  1.97073606
#> [17,] 1.4378102 -0.01228054 1.996398  1.24862640
#> [18,] 1.3448051  0.11766939 2.090327  1.53851017
#> [19,] 2.0835157  0.46585643 2.429864  1.76941771
#> [20,] 2.2976851  0.59886193 2.745862  1.96229187
#> [21,] 1.4933818  0.20960222 2.273323  1.69815242
#> [22,] 2.0977544  0.45602371 2.516474  1.82157870
#> [23,] 1.9404222  0.04251143 2.613838  1.39875339
#> [24,] 1.7520993  0.26621319 2.126168  1.30665014
#> [25,] 2.0770073  0.42114082 2.664408  2.00220423
#> [26,] 1.2213486 -0.07017359 1.871175  1.18200570
#> [27,] 2.1337533  0.40958382 2.650000  1.89486125
#> [28,] 0.5719876 -0.61734502 1.383463  0.51892079
#> [29,] 2.0239231  0.43225107 2.353550  1.70907318
#> [30,] 1.0276153  0.04983927 1.922267  1.37436745
#> [31,] 1.7010002  0.23054876 2.079618  1.23785835
#> [32,] 1.4442184  0.09238470 2.126297  1.61577242
#> [33,] 1.3692581 -0.20804364 2.440725  1.11338531
#> [34,] 1.7708714  0.14892058 2.728350  1.67838386
#> [35,] 1.8381232  0.29330548 2.543737  1.94145423
#> [36,] 1.4637749  0.12893157 2.419612  1.46115704
#> [37,] 0.5921386 -0.62900600 1.473638  0.49974982
#> [38,] 2.2115727  0.57629483 2.641760  1.87937599
#> [39,] 0.6380736 -0.83468225 1.356568  0.22160619
#> [40,] 1.9996620  0.43457617 2.432965  1.78594120
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1371060 0.10474745
#>  [2,] 0.1327910 0.10977642
#>  [3,] 0.1425321 0.10009111
#>  [4,] 0.1276668 0.11882774
#>  [5,] 0.1287273 0.11333021
#>  [6,] 0.1504177 0.09534201
#>  [7,] 0.1247550 0.10805069
#>  [8,] 0.1375235 0.05713516
#>  [9,] 0.1329827 0.11598424
#> [10,] 0.1379630 0.09708237
#> [11,] 0.1294738 0.03037876
#> [12,] 0.1459924 0.07979912
#> [13,] 0.1304450 0.05963002
#> [14,] 0.1285980 0.11032704
#> [15,] 0.1441663 0.10629232
#> [16,] 0.1297769 0.13264709
#> [17,] 0.1485403 0.09490919
#> [18,] 0.1465150 0.09085098
#> [19,] 0.1305868 0.12692366
#> [20,] 0.1282429 0.13448152
#> [21,] 0.1426165 0.10079205
#> [22,] 0.1348256 0.12578362
#> [23,] 0.1394596 0.11405543
#> [24,] 0.1276763 0.11406817
#> [25,] 0.1424907 0.12401160
#> [26,] 0.1455354 0.08497812
#> [27,] 0.1395509 0.12603587
#> [28,] 0.1402255 0.05439391
#> [29,] 0.1345865 0.12249116
#> [30,] 0.1367902 0.07667542
#> [31,] 0.1282056 0.10956058
#> [32,] 0.1527593 0.09763860
#> [33,] 0.1455999 0.08374201
#> [34,] 0.1417710 0.10862421
#> [35,] 0.1453022 0.11788754
#> [36,] 0.1396144 0.09941715
#> [37,] 0.1411355 0.04889239
#> [38,] 0.1265113 0.13304166
#> [39,] 0.1363816 0.04338365
#> [40,] 0.1375566 0.12179921
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5080889  0.5080889
#>  [2,] -0.5057534  0.5057534
#>  [3,]  0.4893913 -0.4893913
#>  [4,] -0.5022098  0.5022098
#>  [5,]  0.4961508 -0.4961508
#>  [6,] -0.5137654  0.5137654
#>  [7,] -0.5041760  0.5041760
#>  [8,] -0.5200863  0.5200863
#>  [9,] -0.5042495  0.5042495
#> [10,]  0.4897813 -0.4897813
#> [11,] -0.5247535  0.5247535
#> [12,]  0.4834577 -0.4834577
#> [13,]  0.4823036 -0.4823036
#> [14,] -0.5045676  0.5045676
#> [15,]  0.4905326 -0.4905326
#> [16,] -0.4992825  0.4992825
#> [17,]  0.4865954 -0.4865954
#> [18,]  0.4860876 -0.4860876
#> [19,]  0.4990842 -0.4990842
#> [20,] -0.4984404  0.4984404
#> [21,]  0.4895454 -0.4895454
#> [22,]  0.4977395 -0.4977395
#> [23,] -0.5063507  0.5063507
#> [24,]  0.4965980 -0.4965980
#> [25,]  0.4953804 -0.4953804
#> [26,]  0.4848653 -0.4848653
#> [27,]  0.4966213 -0.4966213
#> [28,]  0.4785553 -0.4785553
#> [29,]  0.4969762 -0.4969762
#> [30,]  0.4849758 -0.4849758
#> [31,]  0.4953389 -0.4953389
#> [32,]  0.4862233 -0.4862233
#> [33,]  0.4845405 -0.4845405
#> [34,]  0.4917141 -0.4917141
#> [35,]  0.4931468 -0.4931468
#> [36,] -0.5100480  0.5100480
#> [37,]  0.4769556 -0.4769556
#> [38,] -0.4983674  0.4983674
#> [39,]  0.4767673 -0.4767673
#> [40,]  0.4960607 -0.4960607
#> 
#> $L
#> [1] 0.6852441 0.7138206 0.7440346 0.7989755 0.7486824 0.6846976
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5072464