Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11         V12         V13         V14
#>  [1,] -0.3193438 -0.221372493 -0.14598233 -0.19054131 -0.02472397 -0.12475176
#>  [2,] -0.2795701 -0.224342528 -0.09705017 -0.21716267 -0.01157198 -0.10266948
#>  [3,] -0.3130531 -0.131891691 -0.03134502  0.01721742 -0.06130133 -0.10181223
#>  [4,] -0.2286416 -0.150778293 -0.11088367 -0.12866574 -0.11704759  0.03156800
#>  [5,] -0.2607290 -0.184370520 -0.14692175 -0.02944709 -0.10975873 -0.11893651
#>  [6,] -0.2907597 -0.051624667 -0.02724186 -0.05646482 -0.08058136 -0.09369342
#>  [7,] -0.3057989 -0.154596581 -0.19219895 -0.02883071  0.02420276 -0.02253389
#>  [8,] -0.3233545 -0.007873794 -0.15342524 -0.13201149 -0.08947666 -0.17669425
#>  [9,] -0.4288704 -0.144618892 -0.09431187 -0.13243898 -0.13276032 -0.05623406
#> [10,] -0.2419681 -0.139192316 -0.07760518 -0.04434104 -0.11414824 -0.02210367
#>               V15         V16          V17          V18         V19         V2
#>  [1,] -0.12897534 -0.13324982 -0.111863609 -0.012796889  0.06341857 -0.2461665
#>  [2,] -0.12444974 -0.20795324 -0.211359897 -0.096202565 -0.08865105 -0.2798534
#>  [3,] -0.11645718 -0.08296098  0.038470648 -0.036691483  0.12541532 -0.2851009
#>  [4,] -0.04488597  0.03068223  0.023143004 -0.001139903  0.12367587 -0.2891922
#>  [5,] -0.08849269  0.04673438 -0.074132175  0.077929709 -0.08906675 -0.3005768
#>  [6,]  0.06199265  0.09251045 -0.010616844 -0.002308180  0.01972140 -0.4051913
#>  [7,] -0.07694547 -0.09578428  0.002990079  0.021541908 -0.11863589 -0.1960301
#>  [8,] -0.08234807 -0.11408917 -0.129224197 -0.194891684 -0.03802337 -0.1702637
#>  [9,] -0.15737777 -0.13155565 -0.121472470  0.048837685 -0.05250055 -0.3269986
#> [10,] -0.01351548 -0.02493093  0.024672721  0.066193396  0.01819459 -0.2622237
#>                V20         V21          V22         V23         V24
#>  [1,]  0.006607262  0.05757213  0.137476622  0.17803105  0.14293257
#>  [2,] -0.064374717 -0.07614274  0.084366407  0.06005871  0.15955428
#>  [3,]  0.013262668  0.07677862  0.145485211  0.04164457  0.01229067
#>  [4,]  0.019965556  0.11563123  0.018582782  0.18616211  0.16591294
#>  [5,]  0.017299057 -0.09620303  0.104924712  0.08971164  0.03517786
#>  [6,]  0.238003120  0.11497034  0.123312820  0.13190409  0.27342141
#>  [7,] -0.009769097  0.07416037  0.027378306  0.02988721  0.09551433
#>  [8,] -0.111846710 -0.05295216 -0.023716443 -0.07134101 -0.04936659
#>  [9,]  0.051407334  0.12286723  0.006506887  0.19024888  0.18641641
#> [10,]  0.011289399  0.09779271  0.026367902  0.13465552  0.09244104
#>                V25         V26         V27          V28           V29
#>  [1,]  0.187839114  0.02210950  0.08641304  0.111119951  0.0034361012
#>  [2,]  0.060475814  0.12994217  0.14383445  0.213907562  0.1621730035
#>  [3,]  0.035268926 -0.03549698 -0.01206525  0.005535356 -0.0198949335
#>  [4,]  0.139717421  0.15654180  0.14068506  0.130023914  0.2780781605
#>  [5,]  0.058692055  0.04340567  0.10066923 -0.039652701  0.0002499309
#>  [6,]  0.098535164  0.15984339  0.24365711  0.199019192  0.2064421339
#>  [7,]  0.159222968  0.17900994  0.12409282  0.132808773  0.1265717032
#>  [8,] -0.005922772  0.06867235 -0.03620731  0.124796781 -0.0088449713
#>  [9,]  0.193038036  0.22065154  0.24586086  0.200246937  0.1310836448
#> [10,]  0.193391401  0.21528292  0.09786125  0.164335663  0.0241212454
#>               V3         V30           V31         V32          V33         V34
#>  [1,] -0.2606159 -0.05694759  0.0262756242 -0.15215775 -0.129009614 -0.08620395
#>  [2,] -0.3614155 -0.04112670 -0.0001441592  0.04499691 -0.110512011 -0.12081244
#>  [3,] -0.2696264  0.09061155 -0.1069748424 -0.03462541  0.050279739 -0.07307915
#>  [4,] -0.3059074  0.04612476 -0.0167099670 -0.09021332 -0.039688024 -0.03006941
#>  [5,] -0.3052478 -0.09331934 -0.0902213165 -0.06605575 -0.171416015 -0.17524380
#>  [6,] -0.3872116  0.16477560  0.0735578637  0.03679635  0.005912343  0.10534316
#>  [7,] -0.3715010  0.09089157  0.1422449180  0.08287676  0.042800848 -0.04868163
#>  [8,] -0.2070329  0.07296784  0.0031724353  0.06520336  0.067701313  0.08773056
#>  [9,] -0.2649152  0.23416037  0.0550416289  0.02707809  0.055345592  0.06626455
#> [10,] -0.2138888  0.01278721  0.1196042712 -0.11972460 -0.091934400 -0.01906241
#>               V35          V36          V37          V38          V39
#>  [1,] -0.15378159  0.040306890 -0.049679856 -0.001959563 -0.138477581
#>  [2,] -0.03423856 -0.203006668 -0.226495830 -0.154175407 -0.103969067
#>  [3,]  0.00794368 -0.024504272 -0.086651021 -0.139992985 -0.008239194
#>  [4,] -0.06140789 -0.010407619  0.018505488  0.037040672 -0.109184342
#>  [5,] -0.14687877 -0.160971079 -0.124336250 -0.065975412 -0.178271516
#>  [6,]  0.09402536  0.089293477 -0.008979333  0.030512548 -0.007179074
#>  [7,] -0.08060400 -0.126015543  0.014879200 -0.049498451 -0.066871089
#>  [8,] -0.02705488  0.001149757 -0.139742777 -0.009792489  0.003539809
#>  [9,]  0.09028687  0.053531987 -0.065590121 -0.120987371 -0.045958797
#> [10,] -0.17072080 -0.165832965 -0.121222443 -0.108687700 -0.036548732
#>               V4          V40         V41         V42         V43         V44
#>  [1,] -0.2658077 -0.152658056 -0.18617518 -0.13008448 -0.11030280 -0.09610852
#>  [2,] -0.2141016 -0.105441136 -0.21497746 -0.06092372 -0.22208116 -0.11771032
#>  [3,] -0.1961438 -0.070982154  0.06503234 -0.10309595  0.02418687 -0.11475897
#>  [4,] -0.3067074  0.008266709 -0.10365353 -0.07495739 -0.14463189 -0.06734913
#>  [5,] -0.1629539 -0.023781334 -0.17310931 -0.03782429 -0.03251751 -0.15582507
#>  [6,] -0.3477298  0.009893543 -0.10885664  0.03705741 -0.04816861 -0.07113312
#>  [7,] -0.1870228 -0.174824799 -0.07705969 -0.11295771 -0.11955146 -0.19214413
#>  [8,] -0.1935968 -0.171465111 -0.14313410 -0.15627389 -0.12291250 -0.16900563
#>  [9,] -0.3991547  0.007703999  0.04726802 -0.15679954 -0.13878189 -0.05157093
#> [10,] -0.1869015 -0.105564190 -0.16307820 -0.04059571 -0.08189005 -0.05187872
#>               V45         V46         V47        V48        V49         V5
#>  [1,] -0.16855759 -0.15981437 -0.26515697 -0.2572129 -0.3811298 -0.2365452
#>  [2,] -0.11180349 -0.19989560 -0.24486417 -0.1610399 -0.1800275 -0.3253400
#>  [3,] -0.11819613 -0.15392240 -0.11287756 -0.1173880 -0.3164271 -0.1279082
#>  [4,] -0.13338711 -0.18430707 -0.26118748 -0.2326989 -0.2586625 -0.2893956
#>  [5,] -0.16643796 -0.18745271 -0.09123205 -0.2072222 -0.2565446 -0.2642207
#>  [6,] -0.13592568 -0.13975711 -0.18211678 -0.1992172 -0.2475066 -0.2417209
#>  [7,] -0.04234540 -0.05814825 -0.17597269 -0.2503140 -0.2898762 -0.2203639
#>  [8,] -0.06326444 -0.19717554 -0.12816953 -0.1216016 -0.1615211 -0.1213531
#>  [9,] -0.19471623 -0.27332416 -0.16348359 -0.2317883 -0.3781148 -0.2700662
#> [10,] -0.13054145 -0.12865355 -0.23225123 -0.2191157 -0.1967436 -0.2263356
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2758866 -0.3753858 -0.2618297 -0.2914276 -0.2737968 -0.3843991
#>  [2,] -0.2947697 -0.2340799 -0.3048695 -0.2526139 -0.3295499 -0.4129603
#>  [3,] -0.2577473 -0.2923110 -0.3767800 -0.3211219 -0.3911425 -0.2471465
#>  [4,] -0.2874129 -0.3891649 -0.3196864 -0.3543251 -0.3569866 -0.2939722
#>  [5,] -0.2635028 -0.1812826 -0.2568107 -0.2822097 -0.2287484 -0.3355991
#>  [6,] -0.4308992 -0.3915204 -0.4784710 -0.3657686 -0.4987249 -0.4950517
#>  [7,] -0.2466200 -0.2927096 -0.3874526 -0.3307228 -0.3945450 -0.2900243
#>  [8,] -0.2995683 -0.3496665 -0.3274164 -0.1899787 -0.2968108 -0.2437311
#>  [9,] -0.4277356 -0.3069820 -0.3358477 -0.3414462 -0.3131471 -0.3593861
#> [10,] -0.2898021 -0.2544108 -0.2040052 -0.2608398 -0.2279218 -0.3328078
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4071060 -0.2625197 -0.2768329 -0.4065931 -0.2577192 -0.2892891
#>  [2,] -0.3375478 -0.2919271 -0.2788701 -0.3261493 -0.1772322 -0.4021345
#>  [3,] -0.2067182 -0.2262037 -0.2378667 -0.2616322 -0.1506331 -0.3332881
#>  [4,] -0.4457214 -0.4408993 -0.3743252 -0.2924189 -0.2245545 -0.3549381
#>  [5,] -0.2575902 -0.2686494 -0.3337024 -0.3112000 -0.1603127 -0.2952250
#>  [6,] -0.4305821 -0.3924656 -0.4905050 -0.3225682 -0.2104584 -0.3934678
#>  [7,] -0.2880653 -0.3820350 -0.3374607 -0.4031524 -0.1187460 -0.2574437
#>  [8,] -0.1883354 -0.2658614 -0.1984518 -0.2288295 -0.2409258 -0.3357387
#>  [9,] -0.4505417 -0.4571056 -0.3438491 -0.3389606 -0.1993888 -0.4758613
#> [10,] -0.2248499 -0.3503399 -0.3206643 -0.3850695 -0.1778133 -0.3874116
#>               V7         V8          V9
#>  [1,] -0.1548254 -0.2960723 -0.22043351
#>  [2,] -0.2881242 -0.2172280 -0.24983025
#>  [3,] -0.1824615 -0.1357488 -0.11136187
#>  [4,] -0.1948053 -0.2453848 -0.12262876
#>  [5,] -0.2100432 -0.1738915 -0.24006993
#>  [6,] -0.2702971 -0.1578386 -0.03888158
#>  [7,] -0.1201767 -0.1817624 -0.15729272
#>  [8,] -0.1846992 -0.1225531 -0.20656756
#>  [9,] -0.2246664 -0.2710721 -0.23295936
#> [10,] -0.2369384 -0.2561848 -0.18671629
#> 
#> $W[[2]]
#>            [,1]        [,2]         [,3]         [,4]       [,5]        [,6]
#> [1,] -0.1369255 -0.06565137 -0.002151596 -0.007864696 0.01997291 -0.02538857
#> [2,]  0.1155788 -0.02162879  0.001553409 -0.059556916 0.14151308 -0.07153580
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.06453381 -0.13196309 -0.04881381 -0.07198858
#> [2,] -0.04007575  0.02172509  0.09455301 -0.02953236
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.868052e-04  1.595664e-03  1.919917e-03  2.029357e-03  0.0020723834
#>  [2,]  2.789881e-05  1.519006e-04  1.391913e-04  1.742294e-04  0.0003268768
#>  [3,]  1.938413e-05  1.775213e-04  1.630523e-04  2.156525e-04  0.0003932312
#>  [4,] -9.256652e-05 -3.237406e-04 -4.240567e-04 -3.515913e-04 -0.0001849298
#>  [5,]  1.269540e-04  6.964366e-04  8.374850e-04  9.144405e-04  0.0009855341
#>  [6,] -3.570332e-05  7.168853e-06 -3.227731e-05  8.216924e-06  0.0001146572
#>  [7,]  3.530010e-05  2.833217e-04  2.877132e-04  3.479244e-04  0.0005285663
#>  [8,]  1.427385e-04  7.797366e-04  9.227851e-04  9.584311e-04  0.0010080260
#>  [9,]  1.958370e-04  1.048077e-03  1.201224e-03  1.279386e-03  0.0014369138
#> [10,]  7.628721e-05  5.023275e-04  5.508903e-04  6.106772e-04  0.0007691738
#>                V14          V15          V16          V17          V18
#>  [1,] 1.638070e-03 0.0014912683 0.0014576247 0.0012459160 0.0010678343
#>  [2,] 4.733416e-04 0.0006173385 0.0007980156 0.0009780249 0.0011667721
#>  [3,] 5.854784e-04 0.0007667540 0.0010286124 0.0013167566 0.0015728680
#>  [4,] 7.102221e-05 0.0003268445 0.0006475980 0.0010559460 0.0014563836
#>  [5,] 8.612402e-04 0.0008688418 0.0009387546 0.0009214877 0.0009087812
#>  [6,] 2.292723e-04 0.0003511876 0.0005583334 0.0008030760 0.0010393683
#>  [7,] 6.940749e-04 0.0008784105 0.0011531061 0.0014380833 0.0016868605
#>  [8,] 8.274478e-04 0.0007913241 0.0008020141 0.0007139009 0.0006333852
#>  [9,] 1.419406e-03 0.0014797650 0.0016709642 0.0017638470 0.0017706449
#> [10,] 8.645786e-04 0.0009931725 0.0012348837 0.0014653011 0.0016527564
#>                V19            V2          V20          V21         V22
#>  [1,] 0.0019443680  3.795258e-04 0.0024236585 0.0023417017 0.001881075
#>  [2,] 0.0011174855  4.107542e-05 0.0011900352 0.0013005776 0.001469451
#>  [3,] 0.0014465381  3.458807e-05 0.0014898277 0.0016572873 0.001942647
#>  [4,] 0.0010144700 -1.055017e-04 0.0008363782 0.0010153376 0.001519145
#>  [5,] 0.0012347251  1.705712e-04 0.0014327383 0.0014342898 0.001303099
#>  [6,] 0.0008489215 -3.575604e-05 0.0007876613 0.0009164972 0.001187483
#>  [7,] 0.0015948269  5.788114e-05 0.0016506199 0.0018085497 0.002067312
#>  [8,] 0.0010082168  1.935508e-04 0.0012240391 0.0011983920 0.001000711
#>  [9,] 0.0021109956  2.623895e-04 0.0024534829 0.0025349798 0.002412816
#> [10,] 0.0017000708  1.076971e-04 0.0018425487 0.0019884243 0.002153086
#>                V23          V24           V25           V26          V27
#>  [1,] 0.0009140380 0.0003947473 -0.0001619194 -0.0001492871 0.0008087252
#>  [2,] 0.0016946393 0.0018349044  0.0019079881  0.0019207966 0.0019050730
#>  [3,] 0.0023244818 0.0025365817  0.0026378560  0.0026413565 0.0025429379
#>  [4,] 0.0021966375 0.0024978867  0.0026964759  0.0026425204 0.0022665935
#>  [5,] 0.0009831638 0.0008087970  0.0005907756  0.0005955429 0.0009906965
#>  [6,] 0.0015342728 0.0016993019  0.0017909968  0.0017711656 0.0015914589
#>  [7,] 0.0023930870 0.0025716179  0.0026272252  0.0026181148 0.0025670151
#>  [8,] 0.0005383008 0.0003454199  0.0001038528  0.0001006948 0.0005757440
#>  [9,] 0.0022018146 0.0021086749  0.0019103163  0.0019344064 0.0023087739
#> [10,] 0.0023318755 0.0024456461  0.0024432211  0.0024529020 0.0025196771
#>               V28         V29            V3          V30          V31
#>  [1,] 0.001945154 0.001807467  2.507811e-04 0.0015230759 0.0008144741
#>  [2,] 0.001780738 0.001647322  5.032184e-05 0.0014489619 0.0014356678
#>  [3,] 0.002295902 0.002121362  6.474565e-05 0.0018553419 0.0018764186
#>  [4,] 0.001708800 0.001556544 -4.455297e-05 0.0012528340 0.0014997287
#>  [5,] 0.001417713 0.001309886  1.205651e-04 0.0011212979 0.0008327715
#>  [6,] 0.001336529 0.001215841 -6.085357e-06 0.0009809156 0.0010437373
#>  [7,] 0.002383475 0.002198295  8.040506e-05 0.0019059252 0.0019005659
#>  [8,] 0.001129765 0.001074698  1.316852e-04 0.0009485940 0.0006342110
#>  [9,] 0.002700244 0.002518217  2.232020e-04 0.0022542527 0.0019355335
#> [10,] 0.002502085 0.002321794  1.156526e-04 0.0020374416 0.0019138564
#>                V32          V33          V34           V35           V36
#>  [1,] 0.0012515181 0.0007135585 7.543304e-05 -0.0008779286 -1.480955e-03
#>  [2,] 0.0012292504 0.0013463837 1.358643e-03  0.0014362252  1.551862e-03
#>  [3,] 0.0015639215 0.0017343147 1.761964e-03  0.0018994514  2.144352e-03
#>  [4,] 0.0010972141 0.0015345258 1.703956e-03  0.0020659483  2.616282e-03
#>  [5,] 0.0009332469 0.0007585991 4.832787e-04  0.0001161356 -7.761432e-05
#>  [6,] 0.0008095950 0.0010119934 1.050897e-03  0.0011970760  1.518767e-03
#>  [7,] 0.0016093459 0.0017624996 1.765778e-03  0.0018613841  2.089938e-03
#>  [8,] 0.0008202381 0.0005727664 2.226548e-04 -0.0002273072 -4.771394e-04
#>  [9,] 0.0018902776 0.0016842172 1.496702e-03  0.0011983237  1.001910e-03
#> [10,] 0.0016913038 0.0017313947 1.662024e-03  0.0016351022  1.735478e-03
#>                 V37          V38          V39            V4          V40
#>  [1,] -0.0007547974 0.0008337685 0.0015601265  4.035374e-04 0.0006698545
#>  [2,]  0.0013020513 0.0008479857 0.0006099929  4.723679e-05 0.0007618061
#>  [3,]  0.0018246899 0.0011217025 0.0007754229  5.789830e-05 0.0010371345
#>  [4,]  0.0021454108 0.0008701537 0.0003228642 -9.279904e-05 0.0009577206
#>  [5,]  0.0001474203 0.0006408513 0.0008594593  1.849073e-04 0.0005511508
#>  [6,]  0.0013161686 0.0006737636 0.0004068892 -2.643747e-05 0.0007142416
#>  [7,]  0.0018136634 0.0011710487 0.0008623780  8.236092e-05 0.0011003387
#>  [8,] -0.0001525875 0.0005166969 0.0008071828  1.989966e-04 0.0004443337
#>  [9,]  0.0010513150 0.0013741578 0.0014489823  3.005654e-04 0.0011414375
#> [10,]  0.0015636371 0.0012142032 0.0010263076  1.366681e-04 0.0010857751
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0007876795 0.0014276247  1.830803e-03  2.009870e-03  2.313103e-03
#>  [2,] 0.0006790093 0.0005447902  3.366176e-04  1.603957e-04 -6.791924e-06
#>  [3,] 0.0009108100 0.0006790859  3.718798e-04  1.381414e-04 -1.047886e-04
#>  [4,] 0.0007101147 0.0002326701 -2.198016e-04 -5.113957e-04 -9.550558e-04
#>  [5,] 0.0005642891 0.0007831484  8.644344e-04  8.645975e-04  9.252476e-04
#>  [6,] 0.0005810492 0.0003197136  6.597114e-05 -8.580087e-05 -3.273387e-04
#>  [7,] 0.0009784125 0.0007689338  4.724843e-04  2.474310e-04  1.447942e-05
#>  [8,] 0.0005287395 0.0008257856  9.626655e-04  1.028954e-03  1.166065e-03
#>  [9,] 0.0011724118 0.0013421706  1.306808e-03  1.181224e-03  1.232646e-03
#> [10,] 0.0009992834 0.0009051648  7.089811e-04  5.387107e-04  3.785495e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.794847e-03  1.167663e-03  7.889294e-04  4.600479e-04  5.960245e-04
#>  [2,]  3.486772e-05  5.312049e-05  3.404896e-05  1.286161e-05  8.360614e-05
#>  [3,] -3.001165e-05  4.042386e-05  3.018317e-05  1.281623e-05  1.034026e-04
#>  [4,] -7.214955e-04 -3.877850e-04 -2.719620e-04 -1.503723e-04 -8.004879e-05
#>  [5,]  7.338124e-04  4.943999e-04  3.335172e-04  1.937084e-04  2.788400e-04
#>  [6,] -2.470656e-04 -1.043354e-04 -6.747524e-05 -3.252753e-05  3.195170e-06
#>  [7,]  6.536927e-05  1.097145e-04  7.661534e-05  4.039050e-05  1.417290e-04
#>  [8,]  9.382696e-04  6.187954e-04  4.124496e-04  2.359324e-04  2.814107e-04
#>  [9,]  1.044253e-03  7.149222e-04  4.965725e-04  2.827172e-04  4.334491e-04
#> [10,]  3.413215e-04  2.800472e-04  1.974657e-04  1.084445e-04  2.129544e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.426514e-04  1.499177e-04  1.218887e-04  6.140688e-05  6.060245e-05
#>  [2,]  3.228623e-05  1.232903e-05  8.339897e-06  1.835125e-05  1.953468e-05
#>  [3,]  3.691876e-05  1.012216e-05  5.620769e-06  2.142488e-05  2.226296e-05
#>  [4,] -1.461669e-05 -4.892117e-05 -4.495918e-05 -1.998841e-06 -2.355534e-06
#>  [5,]  6.944901e-05  6.638867e-05  5.285445e-05  3.216075e-05  3.176705e-05
#>  [6,]  6.923191e-06 -1.446495e-05 -1.606324e-05  4.737449e-06  4.763328e-06
#>  [7,]  4.528357e-05  1.966159e-05  1.187035e-05  2.475819e-05  2.574356e-05
#>  [8,]  7.775264e-05  7.759832e-05  6.033407e-05  3.049741e-05  3.214008e-05
#>  [9,]  1.139508e-04  1.058395e-04  8.299665e-05  5.541609e-05  5.685376e-05
#> [10,]  6.215066e-05  4.156984e-05  3.060000e-05  3.240695e-05  3.326598e-05
#>                V55          V56          V57           V58           V59
#>  [1,] 5.025679e-05 3.237282e-05 9.822096e-06  5.860731e-05  5.421720e-05
#>  [2,] 2.096312e-05 1.377978e-05 1.848506e-05  1.005593e-05  1.271518e-05
#>  [3,] 2.423113e-05 1.761517e-05 2.552634e-05  1.053982e-05  1.378630e-05
#>  [4,] 4.667546e-06 5.443420e-06 2.072369e-05 -1.224667e-05 -6.753582e-06
#>  [5,] 2.803111e-05 1.823881e-05 1.151049e-05  2.770131e-05  2.692927e-05
#>  [6,] 8.272153e-06 8.051388e-06 1.386667e-05 -1.531897e-06  4.671166e-07
#>  [7,] 2.731069e-05 1.958312e-05 2.686705e-05  1.374194e-05  1.689932e-05
#>  [8,] 2.690591e-05 1.745251e-05 7.912377e-06  2.953067e-05  2.795724e-05
#>  [9,] 5.244915e-05 3.379692e-05 2.937179e-05  4.542189e-05  4.468215e-05
#> [10,] 3.235264e-05 2.352960e-05 2.709034e-05  2.177979e-05  2.378814e-05
#>                 V6          V60           V7            V8            V9
#>  [1,] 0.0004697116 1.826952e-05 0.0004545956  7.573186e-04  0.0013173261
#>  [2,] 0.0001825078 1.541741e-05 0.0002665348  1.950520e-04  0.0001323171
#>  [3,] 0.0002493427 1.959874e-05 0.0003557331  2.459971e-04  0.0001458205
#>  [4,] 0.0001088323 1.164409e-05 0.0002379015 -2.762222e-06 -0.0003315597
#>  [5,] 0.0002603791 1.323886e-05 0.0002828352  3.727908e-04  0.0005681937
#>  [6,] 0.0001240345 8.737907e-06 0.0002082794  1.076541e-04 -0.0000206379
#>  [7,] 0.0002858172 2.083426e-05 0.0003920927  3.006776e-04  0.0002262890
#>  [8,] 0.0002362181 1.007573e-05 0.0002480579  4.104933e-04  0.0006608450
#>  [9,] 0.0004935307 2.876481e-05 0.0005439632  6.624323e-04  0.0009069862
#> [10,] 0.0003300400 2.204881e-05 0.0004181077  3.893812e-04  0.0004238747
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.03126863  0.02878343  0.02030609  0.02925068  0.02223232  0.02125774
#> [2,] -0.03126863 -0.02878343 -0.02030609 -0.02925068 -0.02223232 -0.02125774
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.02677824  0.02326635  0.03289905  0.02422414
#> [2,] -0.02677824 -0.02326635 -0.03289905 -0.02422414
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.24619091 0.24612509 0.31257147 0.61443053 0.29361137 0.72318105
#>  [7] 0.28464960 0.07684619 0.64701605 0.30694232
#> 
#> $B[[2]]
#> [1] 0.005821245 0.017056282
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.001775570 0.002713507 0.003561065 0.003028157 0.001684421 0.002180112
#>  [7] 0.003687465 0.001155735 0.003813859 0.003656839
#> 
#> $vB[[2]]
#> [1]  0.01372796 -0.01372796
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#>  [2,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#>  [3,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>  [4,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#>  [5,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#>  [6,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#>  [7,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#>  [8,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>  [9,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [10,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [11,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [12,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [13,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [14,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [15,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [16,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [17,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [18,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [19,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [20,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [21,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [22,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [23,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [24,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [25,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [26,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [27,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [28,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [29,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [30,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#> [31,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [32,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [33,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [34,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [35,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [36,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [37,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [38,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [39,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [40,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#>  [2,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#>  [3,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>  [4,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#>  [5,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#>  [6,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#>  [7,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#>  [8,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>  [9,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [10,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [11,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [12,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [13,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [14,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [15,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [16,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [17,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [18,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [19,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [20,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [21,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [22,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [23,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [24,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [25,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [26,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [27,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [28,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [29,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [30,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#> [31,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [32,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [33,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [34,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [35,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [36,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [37,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [38,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [39,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [40,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#>  [2,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#>  [3,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>  [4,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#>  [5,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#>  [6,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#>  [7,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#>  [8,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>  [9,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [10,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [11,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [12,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [13,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [14,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [15,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [16,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [17,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [18,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [19,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [20,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [21,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [22,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [23,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [24,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [25,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [26,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [27,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [28,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [29,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [30,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#> [31,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [32,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [33,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [34,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [35,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [36,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [37,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [38,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [39,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [40,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#>  [2,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#>  [3,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>  [4,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#>  [5,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#>  [6,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#>  [7,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#>  [8,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>  [9,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [10,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [11,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [12,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [13,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [14,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [15,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [16,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [17,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [18,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [19,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [20,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [21,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [22,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [23,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [24,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [25,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [26,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [27,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [28,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [29,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [30,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#> [31,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [32,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [33,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [34,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [35,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [36,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [37,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [38,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [39,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [40,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#>  [2,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#>  [3,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>  [4,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#>  [5,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#>  [6,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#>  [7,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#>  [8,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>  [9,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [10,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [11,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [12,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [13,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [14,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [15,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [16,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [17,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [18,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [19,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [20,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [21,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [22,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [23,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [24,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [25,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [26,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [27,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [28,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [29,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [30,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#> [31,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [32,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [33,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [34,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [35,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [36,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [37,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [38,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [39,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [40,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#>  [2,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#>  [3,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#>  [4,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#>  [5,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#>  [6,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#>  [7,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#>  [8,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#>  [9,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [10,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [11,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [12,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [13,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [14,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [15,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [16,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [17,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [18,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [19,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [20,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [21,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [22,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [23,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [24,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [25,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [26,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [27,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [28,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [29,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [30,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#> [31,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [32,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [33,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [34,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [35,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [36,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [37,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [38,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [39,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [40,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4624163 0.3772361 0.5190256 0.7341657 0.3644592 0.8540758 0.5958128
#>  [2,] 0.5192085 0.4060388 0.5604635 0.7591846 0.4380280 0.8519421 0.5813848
#>  [3,] 0.4235077 0.4008274 0.4930477 0.7174006 0.3249253 0.8650263 0.5824963
#>  [4,] 0.4006228 0.3103656 0.4713937 0.7636083 0.2634410 0.9055604 0.6279345
#>  [5,] 0.4768966 0.5125692 0.5018190 0.7473177 0.3959411 0.8631033 0.6633241
#>  [6,] 0.5040737 0.4256310 0.5383054 0.7433891 0.3861545 0.8708193 0.6229291
#>  [7,] 0.3425172 0.2249375 0.4737600 0.6172843 0.3467736 0.7494135 0.4497860
#>  [8,] 0.3632005 0.3532561 0.4650678 0.6409960 0.3552726 0.7750131 0.5243451
#>  [9,] 0.2882056 0.3032806 0.4248059 0.6019268 0.2223812 0.8243532 0.5391981
#> [10,] 0.4128583 0.3829737 0.4868162 0.7025265 0.3369375 0.8523888 0.5720835
#> [11,] 0.2465073 0.1799202 0.3932643 0.5431132 0.2013093 0.7685514 0.3999804
#> [12,] 0.5321568 0.5490251 0.5280270 0.7493953 0.4709264 0.8407926 0.6580199
#> [13,] 0.5902887 0.5294619 0.5639013 0.8191432 0.4210035 0.9177200 0.7029373
#> [14,] 0.5473838 0.4584946 0.5518470 0.8102557 0.4361756 0.9021775 0.6555898
#> [15,] 0.5956624 0.5567392 0.5620098 0.8175193 0.4336471 0.9107397 0.7149058
#> [16,] 0.5824804 0.5477006 0.5590632 0.8212496 0.4109452 0.9186001 0.7194459
#> [17,] 0.2928252 0.2818661 0.4068177 0.6040200 0.2490910 0.8003867 0.5095057
#> [18,] 0.4466180 0.4883561 0.5071158 0.7313609 0.3850098 0.8559287 0.6389126
#> [19,] 0.5431389 0.4916687 0.5327143 0.7537544 0.4370811 0.8543484 0.6039474
#> [20,] 0.3505672 0.3698524 0.4241952 0.6808735 0.2592602 0.8559099 0.5989521
#> [21,] 0.4176101 0.3200693 0.4695627 0.7289696 0.3475948 0.8581871 0.5444491
#> [22,] 0.5043448 0.3938215 0.5163074 0.7673436 0.3317193 0.8949331 0.6058853
#> [23,] 0.3181324 0.2903975 0.4482703 0.6184243 0.2458476 0.8058836 0.4940132
#> [24,] 0.5797928 0.4288252 0.5856750 0.7682093 0.4713427 0.8557613 0.5980836
#> [25,] 0.2277343 0.2379870 0.3704598 0.5014614 0.2117680 0.7323044 0.4345867
#> [26,] 0.5293405 0.4038453 0.5571036 0.7866735 0.4130354 0.8911689 0.5984288
#> [27,] 0.5399523 0.4812760 0.5345586 0.8100675 0.4290655 0.9055460 0.6807625
#> [28,] 0.4376247 0.3122842 0.4666700 0.7323308 0.3586068 0.8517182 0.5505821
#> [29,] 0.4628733 0.3804710 0.5446769 0.7561264 0.3800277 0.8771531 0.5839036
#> [30,] 0.4233008 0.3101846 0.5197491 0.7319710 0.3751451 0.8561235 0.5834892
#> [31,] 0.4772768 0.3628314 0.5645686 0.7380962 0.3977855 0.8556021 0.5751044
#> [32,] 0.4274404 0.3451100 0.4989141 0.7185496 0.2841845 0.8813706 0.5600080
#> [33,] 0.5657139 0.5511831 0.5416569 0.8019633 0.4505663 0.8936684 0.7020765
#> [34,] 0.3973741 0.2812080 0.4854836 0.7130069 0.3522025 0.8463332 0.5221328
#> [35,] 0.3183304 0.2559153 0.3907029 0.5661783 0.2727617 0.6964090 0.4244658
#> [36,] 0.2966023 0.3065321 0.4046130 0.6330379 0.2220695 0.8387563 0.5560748
#> [37,] 0.4729027 0.4520284 0.5079803 0.7412424 0.3599371 0.8769745 0.6092314
#> [38,] 0.5338240 0.5553183 0.5253623 0.7522408 0.4761066 0.8412285 0.6653270
#> [39,] 0.4278334 0.3841946 0.4903632 0.6677276 0.3849720 0.7983717 0.5558749
#> [40,] 0.4148515 0.2642483 0.5304280 0.6685297 0.3632791 0.8114544 0.5008577
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3171992 0.7428429 0.5735133
#>  [2,] 0.2939121 0.7589346 0.6386013
#>  [3,] 0.2847812 0.7549013 0.5843307
#>  [4,] 0.3171168 0.8238587 0.5371925
#>  [5,] 0.3875058 0.8054487 0.6288143
#>  [6,] 0.3249141 0.8139629 0.6017018
#>  [7,] 0.2020589 0.5079468 0.4992506
#>  [8,] 0.3259814 0.6574298 0.5168279
#>  [9,] 0.2626123 0.7004079 0.4428582
#> [10,] 0.3304436 0.7729672 0.5396890
#> [11,] 0.2615280 0.5527261 0.3454486
#> [12,] 0.4215698 0.8037626 0.6477742
#> [13,] 0.4099967 0.8716761 0.6697678
#> [14,] 0.2856904 0.8197347 0.6858921
#> [15,] 0.4311379 0.8781404 0.6760203
#> [16,] 0.4290310 0.8849511 0.6705740
#> [17,] 0.2489673 0.6380911 0.4556497
#> [18,] 0.3807252 0.7916249 0.6028586
#> [19,] 0.3744423 0.7807220 0.6165088
#> [20,] 0.3450269 0.7856986 0.4999752
#> [21,] 0.2433370 0.7361041 0.5589091
#> [22,] 0.3331904 0.8229007 0.5651942
#> [23,] 0.3674963 0.7159681 0.3952815
#> [24,] 0.3191482 0.7651872 0.6355152
#> [25,] 0.2484603 0.5604147 0.3856127
#> [26,] 0.2735652 0.7985173 0.6444544
#> [27,] 0.3175984 0.8312363 0.6895139
#> [28,] 0.2421624 0.7165434 0.5617374
#> [29,] 0.3157895 0.7633647 0.5964283
#> [30,] 0.2478963 0.7186765 0.6173479
#> [31,] 0.3080923 0.7322782 0.5986302
#> [32,] 0.2875536 0.7803614 0.5421986
#> [33,] 0.3978005 0.8430313 0.6913095
#> [34,] 0.2073900 0.6620179 0.5829077
#> [35,] 0.2409319 0.5539190 0.4301513
#> [36,] 0.3160454 0.7516984 0.4418611
#> [37,] 0.4041372 0.8233765 0.5789518
#> [38,] 0.4324671 0.8068950 0.6548283
#> [39,] 0.2972466 0.6791779 0.5730321
#> [40,] 0.2613422 0.6260813 0.5308380
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5025847 0.4974153
#>  [2,] 0.4998961 0.5001039
#>  [3,] 0.5053692 0.4946308
#>  [4,] 0.5073622 0.4926378
#>  [5,] 0.4996288 0.5003712
#>  [6,] 0.4998090 0.5001910
#>  [7,] 0.5086119 0.4913881
#>  [8,] 0.5033917 0.4966083
#>  [9,] 0.5098508 0.4901492
#> [10,] 0.5031196 0.4968804
#> [11,] 0.5104054 0.4895946
#> [12,] 0.4945977 0.5054023
#> [13,] 0.4960428 0.5039572
#> [14,] 0.5008641 0.4991359
#> [15,] 0.4948243 0.5051757
#> [16,] 0.4960561 0.5039439
#> [17,] 0.5097819 0.4902181
#> [18,] 0.5008789 0.4991211
#> [19,] 0.4962668 0.5037332
#> [20,] 0.5059322 0.4940678
#> [21,] 0.5060492 0.4939508
#> [22,] 0.5011550 0.4988450
#> [23,] 0.5046851 0.4953149
#> [24,] 0.4963396 0.5036604
#> [25,] 0.5096439 0.4903561
#> [26,] 0.5013951 0.4986049
#> [27,] 0.5004519 0.4995481
#> [28,] 0.5052424 0.4947576
#> [29,] 0.5032147 0.4967853
#> [30,] 0.5063020 0.4936980
#> [31,] 0.5020862 0.4979138
#> [32,] 0.5054321 0.4945679
#> [33,] 0.4964104 0.5035896
#> [34,] 0.5085070 0.4914930
#> [35,] 0.5057225 0.4942775
#> [36,] 0.5081549 0.4918451
#> [37,] 0.4992428 0.5007572
#> [38,] 0.4942146 0.5057854
#> [39,] 0.5020039 0.4979961
#> [40,] 0.5048419 0.4951581
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]         [,3]        [,4]       [,5]      [,6]
#>  [1,] -0.15061910 -0.50129623  0.076139169 1.015861636 -0.5560618 1.7669326
#>  [2,]  0.07687192 -0.38036512  0.243043377 1.148214514 -0.2491690 1.7499148
#>  [3,] -0.30839044 -0.40201891 -0.027810926 0.931604259 -0.7312279 1.8576796
#>  [4,] -0.40287063 -0.79841080 -0.114550116 1.172564678 -1.0281597 2.2605931
#>  [5,] -0.09247959  0.05028744  0.007275975 1.084357223 -0.4224062 1.8413075
#>  [6,]  0.01629501 -0.29969921  0.153522466 1.063658613 -0.4635056 1.9082223
#>  [7,] -0.65209690 -1.23712112 -0.105056520 0.478037119 -0.6332518 1.0954866
#>  [8,] -0.56150001 -0.60475709 -0.139956765 0.579689796 -0.5959422 1.2368378
#>  [9,] -0.90411479 -0.83172419 -0.303075165 0.413499775 -1.2518437 1.5461237
#> [10,] -0.35216193 -0.47694522 -0.052747316 0.859357744 -0.6769718 1.7534611
#> [11,] -1.11732778 -1.51688816 -0.433611287 0.172882253 -1.3781315 1.2001494
#> [12,]  0.12880501  0.19673251  0.112225579 1.095389572 -0.1164256 1.6641370
#> [13,]  0.36515910  0.11798441  0.257010802 1.510553632 -0.3186554 2.4117647
#> [14,]  0.19010555 -0.16640448  0.208136190 1.451672412 -0.2566980 2.2216568
#> [15,]  0.38742380  0.22793849  0.249322611 1.499629980 -0.2669865 2.3227004
#> [16,]  0.33296424  0.19138431  0.237361115 1.524836978 -0.3600595 2.4234770
#> [17,] -0.88170226 -0.93522403 -0.377136751 0.422243550 -1.1034661 1.3887128
#> [18,] -0.21434476 -0.04658390  0.028464948 1.001538159 -0.4683374 1.7818790
#> [19,]  0.17298582 -0.03332813  0.131044408 1.118737372 -0.2530169 1.7691218
#> [20,] -0.61654717 -0.53285012 -0.305574863 0.757788766 -1.0498174 1.7817266
#> [21,] -0.33259174 -0.75345354 -0.121899798 0.989401016 -0.6296283 1.8003132
#> [22,]  0.01737972 -0.43127678  0.065252681 1.193371741 -0.7004192 2.1421514
#> [23,] -0.76236790 -0.89345413 -0.207661986 0.482865336 -1.1208827 1.4234814
#> [24,]  0.32192294 -0.28664579  0.346114207 1.198227599 -0.1147549 1.7805221
#> [25,] -1.22114901 -1.16374763 -0.530244668 0.005845807 -1.3143010 1.0063457
#> [26,]  0.11749712 -0.38946832  0.229415426 1.304989215 -0.3514314 2.1027364
#> [27,]  0.16015079 -0.07493088  0.138455074 1.450448702 -0.2856650 2.2604251
#> [28,] -0.25080761 -0.78946218 -0.133517818 1.006480551 -0.5814160 1.7481411
#> [29,] -0.14878077 -0.48754936  0.179185497 1.131558722 -0.4894308 1.9657424
#> [30,] -0.30923781 -0.79925656  0.079037440 1.004645635 -0.5102065 1.7834591
#> [31,] -0.09095538 -0.56309625  0.259724436 1.036096853 -0.4147008 1.7792330
#> [32,] -0.29230209 -0.64060418 -0.004343556 0.937278649 -0.9237987 2.0054743
#> [33,]  0.26438502  0.20545215  0.167014890 1.398610595 -0.1983829 2.1287719
#> [34,] -0.41641848 -0.93847746 -0.058081894 0.910032836 -0.6093717 1.7061262
#> [35,] -0.76145537 -1.06730830 -0.444358413 0.266275499 -0.9806558 0.8302556
#> [36,] -0.86353018 -0.81638263 -0.386280575 0.545271852 -1.2536466 1.6490035
#> [37,] -0.10849543 -0.19247836  0.031923937 1.052436003 -0.5756370 1.9640865
#> [38,]  0.13550276  0.22218278  0.101536221 1.110599442 -0.0956466 1.6673972
#> [39,] -0.29069623 -0.47178175 -0.038552026 0.697925142 -0.4684973 1.3761485
#> [40,] -0.34394504 -1.02400332  0.121862625 0.701542719 -0.5611600 1.4594881
#>               [,7]       [,8]       [,9]         [,10]
#>  [1,]  0.388048559 -0.7666730 1.06079744  2.961998e-01
#>  [2,]  0.328460799 -0.8764591 1.14684721  5.692987e-01
#>  [3,]  0.333029323 -0.9208675 1.12492587  3.405771e-01
#>  [4,]  0.523366107 -0.7670535 1.54271267  1.490455e-01
#>  [5,]  0.678142774 -0.4578087 1.42070384  5.271335e-01
#>  [6,]  0.501999577 -0.7312792 1.47596890  4.125611e-01
#>  [7,] -0.201535294 -1.3734759 0.03178985 -2.997697e-03
#>  [8,]  0.097457299 -0.7264175 0.65186134  6.733697e-02
#>  [9,]  0.157114855 -1.0324352 0.84924108 -2.295703e-01
#> [10,]  0.290356835 -0.7061794 1.22514217  1.590909e-01
#> [11,] -0.405546663 -1.0380418 0.21169144 -6.391061e-01
#> [12,]  0.654482678 -0.3163325 1.40997849  6.092696e-01
#> [13,]  0.861324347 -0.3639788 1.91586021  7.071352e-01
#> [14,]  0.643701733 -0.9164076 1.51455086  7.809837e-01
#> [15,]  0.919331051 -0.2772101 1.97493658  7.355425e-01
#> [16,]  0.941714850 -0.2858057 2.04017538  7.107825e-01
#> [17,]  0.038027469 -1.1041277 0.56708875 -1.778686e-01
#> [18,]  0.570647531 -0.4864712 1.33474781  4.173901e-01
#> [19,]  0.421939910 -0.5132058 1.26987871  4.747558e-01
#> [20,]  0.401100835 -0.6409718 1.29918986 -9.914077e-05
#> [21,]  0.178267025 -1.1344706 1.02581668  2.367357e-01
#> [22,]  0.430048546 -0.6937905 1.53612528  2.622699e-01
#> [23,] -0.023948232 -0.5429729 0.92454913 -4.251648e-01
#> [24,]  0.397486287 -0.7576891 1.18133172  5.559512e-01
#> [25,] -0.263161669 -1.1068409 0.24284528 -4.657921e-01
#> [26,]  0.398922527 -0.9766088 1.37705333  5.947505e-01
#> [27,]  0.757278023 -0.7648306 1.59441433  7.978475e-01
#> [28,]  0.203023047 -1.1408607 0.92737967  2.482163e-01
#> [29,]  0.338818822 -0.7731897 1.17121580  3.906049e-01
#> [30,]  0.337113421 -1.1098639 0.93790585  4.783065e-01
#> [31,]  0.302708273 -0.8090533 1.00621224  3.997608e-01
#> [32,]  0.241194721 -0.9072954 1.26777365  1.691969e-01
#> [33,]  0.857205680 -0.4146381 1.68095779  8.062485e-01
#> [34,]  0.088589142 -1.3407299 0.67229986  3.347211e-01
#> [35,] -0.304467262 -1.1475773 0.21651811 -2.812340e-01
#> [36,]  0.225246740 -0.7720055 1.10769103 -2.336121e-01
#> [37,]  0.444082700 -0.3882559 1.53939291  3.184719e-01
#> [38,]  0.687124905 -0.2717925 1.42995944  6.403311e-01
#> [39,]  0.224437188 -0.8604440 0.74999618  2.942329e-01
#> [40,]  0.003430846 -1.0390039 0.51544184  1.235087e-01
#> 
#> $pre[[3]]
#>              [,1]        [,2]
#>  [1,] -0.10242812 -0.11276718
#>  [2,] -0.10537839 -0.10496295
#>  [3,] -0.09841293 -0.11989043
#>  [4,] -0.09890341 -0.12835427
#>  [5,] -0.12151601 -0.12003126
#>  [6,] -0.11156061 -0.11079645
#>  [7,] -0.06246636 -0.09691717
#>  [8,] -0.08828383 -0.10185092
#>  [9,] -0.07672025 -0.11612872
#> [10,] -0.09922701 -0.11170557
#> [11,] -0.05787860 -0.09950614
#> [12,] -0.12930269 -0.10769254
#> [13,] -0.13757462 -0.12174555
#> [14,] -0.11477357 -0.11823015
#> [15,] -0.14180603 -0.12110234
#> [16,] -0.14083501 -0.12505897
#> [17,] -0.07248934 -0.11162198
#> [18,] -0.11498255 -0.11849835
#> [19,] -0.12063035 -0.10569742
#> [20,] -0.09795994 -0.12168969
#> [21,] -0.08655161 -0.11074965
#> [22,] -0.11135523 -0.11597522
#> [23,] -0.08523148 -0.10397227
#> [24,] -0.11390430 -0.09926252
#> [25,] -0.06067647 -0.09925698
#> [26,] -0.10639940 -0.11197979
#> [27,] -0.12023492 -0.12204256
#> [28,] -0.08779608 -0.10876646
#> [29,] -0.10191684 -0.11477567
#> [30,] -0.08912535 -0.11433449
#> [31,] -0.10041587 -0.10876074
#> [32,] -0.09627525 -0.11800457
#> [33,] -0.13441799 -0.12005923
#> [34,] -0.07800704 -0.11203849
#> [35,] -0.06725903 -0.09014985
#> [36,] -0.08471504 -0.11733741
#> [37,] -0.11812609 -0.11509732
#> [38,] -0.13132873 -0.10818592
#> [39,] -0.09516324 -0.10317877
#> [40,] -0.08075423 -0.10012255
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5025847  0.5025847
#>  [2,] -0.4998961  0.4998961
#>  [3,]  0.4946308 -0.4946308
#>  [4,] -0.5073622  0.5073622
#>  [5,]  0.5003712 -0.5003712
#>  [6,] -0.4998090  0.4998090
#>  [7,]  0.4913881 -0.4913881
#>  [8,]  0.4966083 -0.4966083
#>  [9,]  0.4901492 -0.4901492
#> [10,] -0.5031196  0.5031196
#> [11,]  0.4895946 -0.4895946
#> [12,] -0.4945977  0.4945977
#> [13,] -0.4960428  0.4960428
#> [14,]  0.4991359 -0.4991359
#> [15,] -0.4948243  0.4948243
#> [16,] -0.4960561  0.4960561
#> [17,]  0.4902181 -0.4902181
#> [18,]  0.4991211 -0.4991211
#> [19,] -0.4962668  0.4962668
#> [20,]  0.4940678 -0.4940678
#> [21,] -0.5060492  0.5060492
#> [22,]  0.4988450 -0.4988450
#> [23,] -0.5046851  0.5046851
#> [24,] -0.4963396  0.4963396
#> [25,]  0.4903561 -0.4903561
#> [26,]  0.4986049 -0.4986049
#> [27,]  0.4995481 -0.4995481
#> [28,] -0.5052424  0.5052424
#> [29,]  0.4967853 -0.4967853
#> [30,] -0.5063020  0.5063020
#> [31,]  0.4979138 -0.4979138
#> [32,]  0.4945679 -0.4945679
#> [33,] -0.4964104  0.4964104
#> [34,]  0.4914930 -0.4914930
#> [35,]  0.4942775 -0.4942775
#> [36,]  0.4918451 -0.4918451
#> [37,] -0.4992428  0.4992428
#> [38,] -0.4942146  0.4942146
#> [39,]  0.4979961 -0.4979961
#> [40,]  0.4951581 -0.4951581
#> 
#> $L
#> [1] 0.7038992 0.8056144 0.7831676 0.7059424 0.8060018 0.6877903
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319