Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10           V11          V12          V13
#>  [1,] -0.3443631 -0.18005531 -0.0739751546  0.001241420  0.018759901
#>  [2,] -0.3758324  0.03576134 -0.0951078149  0.072336502 -0.008965035
#>  [3,] -0.4188985 -0.10164412 -0.0001905203 -0.145828821 -0.104284264
#>  [4,] -0.3386380 -0.03374565 -0.1772516036 -0.134572099 -0.114338616
#>  [5,] -0.1875858 -0.21715978 -0.1556360488 -0.031066396 -0.155608329
#>  [6,] -0.2077874 -0.18015515 -0.0412868642 -0.167640837 -0.048324004
#>  [7,] -0.2763197 -0.07657333 -0.1836326552 -0.100903450 -0.064656493
#>  [8,] -0.3170833 -0.11996588 -0.0179648424 -0.122460746  0.083294240
#>  [9,] -0.2184910 -0.10804917 -0.0381397615 -0.009064154 -0.111716712
#> [10,] -0.1647032 -0.16252003 -0.0996471706 -0.136208552 -0.124869916
#>                 V14          V15          V16          V17           V18
#>  [1,] -0.0438917409  0.072292206  0.049289349  0.006657440  0.0877637708
#>  [2,]  0.0241499881  0.027537086 -0.001499713  0.082669399  0.1450312514
#>  [3,] -0.0392536416  0.092494239  0.063414462  0.173169974  0.1706980001
#>  [4,]  0.0200142631 -0.007373295 -0.106905522 -0.025846715 -0.0009668378
#>  [5,] -0.1892304150 -0.207015966 -0.106827231 -0.096316210 -0.1179138117
#>  [6,] -0.0027595456  0.016575286  0.074076745  0.061626408  0.0561061295
#>  [7,] -0.0590028374 -0.138505428  0.026264134 -0.086727437 -0.1379919267
#>  [8,]  0.0578391786 -0.034181906  0.036920853 -0.055062782  0.0449378754
#>  [9,]  0.0002543268 -0.023348401  0.058972737 -0.007116347  0.0792639273
#> [10,] -0.1290018137 -0.052327533  0.005333865 -0.058643712 -0.0042998620
#>               V19         V2         V20         V21        V22         V23
#>  [1,]  0.03674748 -0.3553785  0.16083282  0.20010777 0.14961625  0.21821256
#>  [2,]  0.19099445 -0.2423582  0.25088663  0.17144955 0.19808141  0.17390310
#>  [3,]  0.12193498 -0.3904796  0.13451024  0.24050160 0.20472011  0.24651028
#>  [4,]  0.18898900 -0.3438846  0.18406053  0.07328498 0.21420819  0.15845890
#>  [5,] -0.08620904 -0.3463735 -0.08000540  0.02770173 0.02420508  0.15144994
#>  [6,]  0.05656660 -0.3705764 -0.01296014  0.04336580 0.14778930  0.13153025
#>  [7,] -0.08505042 -0.1937867  0.03667513  0.05431354 0.09982865  0.08208645
#>  [8,]  0.16642539 -0.2467270  0.14770136  0.05570924 0.23052338  0.14352862
#>  [9,]  0.12863779 -0.2377394  0.13640799  0.23869845 0.12162405  0.15530296
#> [10,]  0.04378016 -0.1725156 -0.03081872 -0.04646990 0.03382256 -0.05317812
#>               V24          V25        V26        V27        V28         V29
#>  [1,]  0.15163680  0.207904115 0.12137531 0.18116425 0.01349710  0.08684381
#>  [2,]  0.19189827  0.068994936 0.18570169 0.12602421 0.13308700  0.06320108
#>  [3,]  0.22796847  0.262379859 0.30319892 0.26108487 0.22754617  0.19187724
#>  [4,]  0.16121835 -0.004379055 0.09002272 0.06424238 0.12175000  0.04909597
#>  [5,]  0.23999810  0.197531422 0.28819636 0.28233981 0.20938864  0.26658165
#>  [6,]  0.12820537  0.162101611 0.04312551 0.14796061 0.02624986 -0.01317836
#>  [7,]  0.03579830  0.051859806 0.09721875 0.05797112 0.12142453 -0.05501665
#>  [8,]  0.15861856  0.239110295 0.22341634 0.20596659 0.10864489  0.17300352
#>  [9,]  0.22279922  0.313093499 0.18786337 0.18612222 0.14322972  0.16110307
#> [10,] -0.05443819 -0.039649957 0.02448669 0.09100741 0.11577835  0.04651812
#>               V3          V30          V31          V32         V33
#>  [1,] -0.3440266  0.062049011  0.093930682 -0.004217913  0.04965729
#>  [2,] -0.2516484 -0.012028134  0.059833641  0.033693677  0.00661240
#>  [3,] -0.3320578  0.208453773  0.140184550  0.018930240 -0.01621351
#>  [4,] -0.3927274  0.005044748 -0.112927277 -0.072691611 -0.17263704
#>  [5,] -0.1953974  0.077900745  0.037914753 -0.011114403  0.02778638
#>  [6,] -0.2860994  0.003711016 -0.007519874  0.026315666  0.01517299
#>  [7,] -0.2642860 -0.090631257  0.035346079  0.097977571  0.07753956
#>  [8,] -0.3212335  0.097280918  0.110091054  0.017335239 -0.02636118
#>  [9,] -0.2823107  0.109099986 -0.024999750 -0.003693681  0.01820497
#> [10,] -0.2751889  0.016308424 -0.134882056 -0.009274883 -0.12722367
#>                V34          V35          V36          V37         V38
#>  [1,] -0.063385449 -0.060011511 -0.089499558 -0.052416802 -0.05807920
#>  [2,] -0.012189182  0.055135246  0.009618537 -0.004754338 -0.03005664
#>  [3,]  0.087106321  0.010305132 -0.046836987  0.012212694  0.01964585
#>  [4,] -0.100285630 -0.136029831 -0.074759314 -0.073562451 -0.08040999
#>  [5,] -0.118797496 -0.200114030 -0.149492480 -0.109489849 -0.07603760
#>  [6,]  0.002507003 -0.190656303 -0.051431431 -0.088611749 -0.06359299
#>  [7,] -0.014104838 -0.010931136 -0.067937876 -0.030706853 -0.13107227
#>  [8,] -0.081649500 -0.090243181 -0.063073647 -0.198171136 -0.09732341
#>  [9,] -0.011076013 -0.034481976 -0.034158867 -0.020402478 -0.14962089
#> [10,] -0.129116564 -0.006398974 -0.019127528 -0.205445654 -0.18260251
#>                V39         V4          V40          V41           V42
#>  [1,] -0.018914772 -0.3183500 -0.056363820 -0.065685402 -0.0885739421
#>  [2,] -0.118236873 -0.3420721 -0.038275079 -0.011577354  0.0028156675
#>  [3,] -0.010627058 -0.3227880 -0.061876963 -0.089043493  0.0107804018
#>  [4,] -0.016670039 -0.2021742 -0.058028529  0.004799659 -0.1706336875
#>  [5,] -0.060391035 -0.3013716 -0.092922688 -0.093413055 -0.1080868868
#>  [6,] -0.108603007 -0.3440527 -0.121746435 -0.136251946 -0.1828289374
#>  [7,] -0.086183714 -0.2149631  0.005190451 -0.167227179  0.0003489547
#>  [8,] -0.119089623 -0.3370347 -0.051999422 -0.034687611 -0.1516698246
#>  [9,] -0.056006742 -0.2524430 -0.041788020 -0.036944623 -0.0222633294
#> [10,]  0.001944762 -0.2715784 -0.023648906 -0.135983989 -0.0537299114
#>                 V43         V44         V45         V46         V47        V48
#>  [1,] -0.0138277617 -0.19121788 -0.14227452 -0.24489417 -0.28910275 -0.2302107
#>  [2,] -0.0004698407 -0.02874700  0.00621256 -0.19367629 -0.26434611 -0.3108453
#>  [3,] -0.0630869252 -0.15377668 -0.18498742 -0.15835797 -0.12509256 -0.2082327
#>  [4,] -0.1126152654 -0.06592249 -0.14967131 -0.18606617 -0.29734096 -0.3049897
#>  [5,] -0.0867318675 -0.05254487 -0.10662118 -0.12251757 -0.13630127 -0.1765881
#>  [6,] -0.2276779389 -0.06103745 -0.03357206 -0.10912665 -0.13143198 -0.1836782
#>  [7,] -0.0188052906 -0.07655616 -0.07893101 -0.07545079 -0.13277216 -0.2654854
#>  [8,] -0.1348872514 -0.08733688 -0.11612188 -0.24839597 -0.13213791 -0.1463206
#>  [9,] -0.0964713711 -0.17752652 -0.18992295 -0.05437095 -0.16317544 -0.1234589
#> [10,] -0.0115102246 -0.19440324 -0.17511496 -0.19513783 -0.07774876 -0.2711240
#>              V49         V5        V50        V51        V52        V53
#>  [1,] -0.2153950 -0.2773251 -0.3791696 -0.2528166 -0.3227564 -0.2692517
#>  [2,] -0.2165899 -0.2069437 -0.3831799 -0.3659462 -0.3764829 -0.3972067
#>  [3,] -0.3263344 -0.2629481 -0.4145410 -0.4241443 -0.4454481 -0.4629739
#>  [4,] -0.3010525 -0.2855250 -0.3067834 -0.2782660 -0.2832757 -0.3507780
#>  [5,] -0.1888556 -0.2124269 -0.2757600 -0.3269623 -0.3350489 -0.2656921
#>  [6,] -0.1922351 -0.2515416 -0.3680404 -0.4051609 -0.2856581 -0.3583081
#>  [7,] -0.3030915 -0.2535366 -0.1875698 -0.3162956 -0.2444943 -0.3391186
#>  [8,] -0.2459008 -0.1384482 -0.2276657 -0.3728363 -0.2826125 -0.2606917
#>  [9,] -0.2895644 -0.2644949 -0.2892758 -0.2919875 -0.2714682 -0.1924602
#> [10,] -0.3165922 -0.2512326 -0.2339273 -0.3297526 -0.2534460 -0.2222111
#>              V54        V55        V56        V57        V58        V59
#>  [1,] -0.3317858 -0.3889099 -0.2857847 -0.3651090 -0.3696609 -0.4021459
#>  [2,] -0.4088650 -0.2715051 -0.4128008 -0.3360306 -0.3795128 -0.2853501
#>  [3,] -0.4183327 -0.4780134 -0.3818457 -0.4010461 -0.5130533 -0.3751454
#>  [4,] -0.2729892 -0.3175211 -0.3802277 -0.3203818 -0.3670471 -0.3422579
#>  [5,] -0.3056182 -0.2848846 -0.3385607 -0.3026880 -0.2488630 -0.3176088
#>  [6,] -0.3343409 -0.2758066 -0.3452510 -0.4059300 -0.2318713 -0.4357901
#>  [7,] -0.3421292 -0.2304335 -0.2442740 -0.3530846 -0.2393065 -0.2125034
#>  [8,] -0.4172308 -0.2892046 -0.3290459 -0.3071830 -0.3467362 -0.3392399
#>  [9,] -0.2519988 -0.2686486 -0.2484386 -0.2265680 -0.2053971 -0.1758272
#> [10,] -0.2944260 -0.3316749 -0.2817201 -0.3007655 -0.2554892 -0.3467263
#>                V6        V60         V7          V8          V9
#>  [1,] -0.22658310 -0.3511166 -0.2685678 -0.23082835 -0.14849495
#>  [2,] -0.22058097 -0.2785863 -0.1231218 -0.17423427 -0.04245266
#>  [3,] -0.15104659 -0.4178797 -0.1889759 -0.17418026 -0.16835092
#>  [4,] -0.17437438 -0.4179613 -0.2093859 -0.10727778 -0.08321426
#>  [5,] -0.12378936 -0.2357496 -0.1708533 -0.12422372 -0.08035053
#>  [6,] -0.27159451 -0.3807511 -0.2844228 -0.07964032 -0.10402011
#>  [7,] -0.14339654 -0.3188879 -0.2080194 -0.18428038 -0.16445644
#>  [8,] -0.20595244 -0.3661686 -0.1446396 -0.24386544 -0.17834553
#>  [9,] -0.22763414 -0.2987146 -0.1509957 -0.18761960 -0.20821150
#> [10,] -0.08138949 -0.3122167 -0.1395445 -0.19260734 -0.11369756
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]        [,4]         [,5]        [,6]
#> [1,] -0.04379200 0.01193159 -0.03753115 -0.04003773 -0.002657036 -0.10778317
#> [2,]  0.01365462 0.04730766 -0.07760973 -0.09860446  0.046247335 -0.02012601
#>            [,7]        [,8]        [,9]       [,10]
#> [1,] 0.02875533 -0.06440001 -0.06231523 0.002765898
#> [2,] 0.09251199  0.05241491 -0.06321727 0.128473396
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12          V13
#>  [1,]  7.407268e-05  2.849505e-04  3.301289e-04  4.144755e-04 4.805335e-04
#>  [2,]  6.604339e-05  2.611084e-04  2.993047e-04  3.774544e-04 4.455454e-04
#>  [3,]  3.597624e-06 -1.612053e-04 -1.839132e-04 -3.551291e-05 1.304161e-04
#>  [4,] -1.355723e-05 -6.632145e-04 -8.017467e-04 -4.532368e-04 5.014259e-05
#>  [5,]  7.719874e-05  2.805188e-04  3.303740e-04  4.279718e-04 5.056728e-04
#>  [6,]  8.401784e-05  4.474431e-04  5.314463e-04  5.637310e-04 5.549448e-04
#>  [7,]  5.658135e-05  2.563851e-04  3.077614e-04  3.561923e-04 3.949840e-04
#>  [8,]  1.191566e-04  7.169198e-04  8.318846e-04  7.825475e-04 6.705484e-04
#>  [9,]  3.892768e-05 -1.983296e-05 -2.761075e-05  1.503381e-04 3.416883e-04
#> [10,]  6.553827e-05  5.455763e-04  6.671436e-04  5.892954e-04 3.978775e-04
#>                V14           V15           V16           V17           V18
#>  [1,] 0.0006844867  0.0008192046  0.0008324135  0.0008472615  0.0009455532
#>  [2,] 0.0006423898  0.0007561782  0.0007567748  0.0007611734  0.0008592479
#>  [3,] 0.0004169379  0.0006229008  0.0006411604  0.0006700297  0.0007483165
#>  [4,] 0.0009830088  0.0017294884  0.0019105579  0.0020133509  0.0020624876
#>  [5,] 0.0007262792  0.0008913695  0.0009242018  0.0009566911  0.0010719467
#>  [6,] 0.0006075986  0.0006350966  0.0006412855  0.0006511454  0.0007554507
#>  [7,] 0.0005031470  0.0005852541  0.0006133292  0.0006317985  0.0007112111
#>  [8,] 0.0005955658  0.0004915869  0.0004471237  0.0004065772  0.0005211668
#>  [9,] 0.0007327459  0.0010265546  0.0010654600  0.0011102071  0.0012061897
#> [10,] 0.0001109984 -0.0001055621 -0.0001593990 -0.0001871216 -0.0001356108
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0010594239  8.370275e-05 0.0011720478 0.0012813721 0.0014465665
#>  [2,] 0.0009638813  7.373026e-05 0.0010703455 0.0011678510 0.0013271149
#>  [3,] 0.0006068967  7.596128e-07 0.0004627575 0.0004471822 0.0007304699
#>  [4,] 0.0014265478 -2.597447e-05 0.0010512265 0.0010873903 0.0018299823
#>  [5,] 0.0011968939  8.663291e-05 0.0013060295 0.0014315998 0.0016207203
#>  [6,] 0.0010007938  9.875247e-05 0.0012066432 0.0013361739 0.0013911151
#>  [7,] 0.0008347374  6.562777e-05 0.0009569004 0.0010534376 0.0011512526
#>  [8,] 0.0009610854  1.404414e-04 0.0013287100 0.0014889050 0.0013716734
#>  [9,] 0.0010931216  4.087051e-05 0.0010079230 0.0010590218 0.0014074561
#> [10,] 0.0002322506  8.203288e-05 0.0004800624 0.0005595831 0.0003623427
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0015883509 0.0015437197 1.394108e-03 0.0012512840 0.0013221592
#>  [2,] 0.0014762661 0.0014417978 1.289970e-03 0.0011651810 0.0012316929
#>  [3,] 0.0010087214 0.0009438473 8.205642e-04 0.0005936282 0.0006311656
#>  [4,] 0.0025432953 0.0024958163 2.455029e-03 0.0019129702 0.0017979633
#>  [5,] 0.0017674958 0.0017025552 1.532053e-03 0.0013593419 0.0014140052
#>  [6,] 0.0014300913 0.0014079448 1.268898e-03 0.0012250695 0.0012981616
#>  [7,] 0.0012392791 0.0012242718 1.122703e-03 0.0010530519 0.0010848369
#>  [8,] 0.0012410216 0.0012155390 1.044276e-03 0.0011024986 0.0012116688
#>  [9,] 0.0017169614 0.0016393877 1.492646e-03 0.0012161805 0.0012621891
#> [10,] 0.0001510725 0.0001617160 8.477851e-05 0.0002153556 0.0003195074
#>                V28           V29            V3           V30           V31
#>  [1,] 0.0014593962  1.510540e-03  9.617059e-05  0.0015284738  0.0012311024
#>  [2,] 0.0013620120  1.434681e-03  8.566512e-05  0.0014431766  0.0011583562
#>  [3,] 0.0008434317  1.060666e-03  9.410824e-06  0.0010451338  0.0009188642
#>  [4,] 0.0022552137  2.917985e-03 -1.352058e-05  0.0030314633  0.0027914516
#>  [5,] 0.0015335645  1.597410e-03  9.805562e-05  0.0016095140  0.0012980273
#>  [6,] 0.0013421753  1.280500e-03  1.085994e-04  0.0012454744  0.0009502640
#>  [7,] 0.0011380183  1.148688e-03  7.292776e-05  0.0011140526  0.0008910079
#>  [8,] 0.0011838255  9.984027e-04  1.502549e-04  0.0010315414  0.0006804941
#>  [9,] 0.0015119120  1.735276e-03  5.512501e-05  0.0017683640  0.0015198829
#> [10,] 0.0002138198 -6.147026e-05  8.556175e-05 -0.0001420915 -0.0002752346
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0008661519  8.439005e-04  0.0007769507  0.0007383594  6.081887e-04
#>  [2,]  0.0008165075  8.099123e-04  0.0007373760  0.0006709076  5.588729e-04
#>  [3,]  0.0004956354  5.101365e-04  0.0005139705  0.0005644765  5.617533e-04
#>  [4,]  0.0016954889  1.659389e-03  0.0017967078  0.0022036647  2.232441e-03
#>  [5,]  0.0009066706  8.832599e-04  0.0008208997  0.0007885938  6.540354e-04
#>  [6,]  0.0007468570  7.141191e-04  0.0006135928  0.0004963819  3.537423e-04
#>  [7,]  0.0006682974  6.402217e-04  0.0005821141  0.0005271591  4.352695e-04
#>  [8,]  0.0006315123  6.240873e-04  0.0004892183  0.0002779810  5.323351e-05
#>  [9,]  0.0009409154  9.244777e-04  0.0009118280  0.0009826347  9.193331e-04
#> [10,] -0.0000624224 -9.806546e-05 -0.0002303379 -0.0004317409 -5.323199e-04
#>                 V37          V38          V39           V4          V40
#>  [1,]  0.0006463893 6.998301e-04 0.0006306213 1.166032e-04 0.0004523750
#>  [2,]  0.0005927578 6.414158e-04 0.0005883046 1.050724e-04 0.0004137652
#>  [3,]  0.0004987168 3.903543e-04 0.0002759948 2.053543e-05 0.0001936661
#>  [4,]  0.0018237825 1.185481e-03 0.0005808620 3.696963e-05 0.0004479802
#>  [5,]  0.0006800114 7.273821e-04 0.0006470202 1.222621e-04 0.0004785234
#>  [6,]  0.0004461210 5.885931e-04 0.0006098548 1.284251e-04 0.0004606636
#>  [7,]  0.0004607751 4.948204e-04 0.0004617217 9.234395e-05 0.0003605946
#>  [8,]  0.0002392112 5.855178e-04 0.0007232817 1.670344e-04 0.0004971245
#>  [9,]  0.0008594895 7.432284e-04 0.0005571112 7.534904e-05 0.0003973690
#> [10,] -0.0002998384 4.012614e-07 0.0002064959 7.976681e-05 0.0001635384
#>                V41           V42           V43           V44           V45
#>  [1,] 0.0004666684  3.862717e-04  3.302379e-04  0.0002450016  0.0002789223
#>  [2,] 0.0004087434  3.328019e-04  2.913390e-04  0.0002181804  0.0002427126
#>  [3,] 0.0001494823 -3.062717e-05 -1.026559e-04 -0.0002257929 -0.0002475487
#>  [4,] 0.0003794238 -2.498595e-04 -4.226514e-04 -0.0007090616 -0.0007686432
#>  [5,] 0.0004955009  3.995950e-04  3.442544e-04  0.0002566578  0.0002840062
#>  [6,] 0.0004919054  5.108083e-04  4.778873e-04  0.0004338790  0.0004645280
#>  [7,] 0.0003777429  3.388405e-04  3.011750e-04  0.0002493605  0.0002607030
#>  [8,] 0.0005414615  6.845204e-04  6.979797e-04  0.0007145159  0.0007997940
#>  [9,] 0.0003679347  1.369257e-04  4.199955e-05 -0.0001080460 -0.0001064482
#> [10,] 0.0002263538  4.630111e-04  4.873213e-04  0.0005377882  0.0005668394
#>                 V46           V47           V48           V49            V5
#>  [1,]  2.402603e-04  2.438765e-04  1.361595e-04  7.043339e-05  1.119987e-04
#>  [2,]  2.089152e-04  2.191697e-04  1.243131e-04  6.356646e-05  9.920501e-05
#>  [3,] -1.399142e-04  3.118530e-05 -4.265056e-05 -3.562526e-05 -3.999037e-06
#>  [4,] -4.794793e-04 -9.810281e-05 -2.550150e-04 -1.741066e-04 -4.252063e-05
#>  [5,]  2.416577e-04  2.496485e-04  1.339331e-04  7.012839e-05  1.183992e-04
#>  [6,]  3.602965e-04  2.862518e-04  1.984027e-04  1.130990e-04  1.461090e-04
#>  [7,]  2.109357e-04  1.896268e-04  1.177071e-04  6.524850e-05  1.017016e-04
#>  [8,]  5.943831e-04  3.945970e-04  3.089962e-04  1.786293e-04  1.931120e-04
#>  [9,] -2.860717e-05  1.254333e-04  7.237544e-06 -1.130505e-05  4.680340e-05
#> [10,]  4.104585e-04  2.441058e-04  2.292309e-04  1.404218e-04  1.226669e-04
#>                V50           V51          V52          V53          V54
#>  [1,] 5.156510e-05  3.146825e-05 2.917814e-05 2.776167e-05 2.050289e-05
#>  [2,] 4.694448e-05  2.925771e-05 2.751754e-05 2.580670e-05 1.860277e-05
#>  [3,] 2.289555e-05  5.707249e-06 1.086496e-05 1.760996e-05 7.280108e-06
#>  [4,] 4.999792e-05 -4.401493e-06 1.983203e-05 3.868956e-05 1.064236e-05
#>  [5,] 5.487164e-05  3.329649e-05 3.074882e-05 2.996194e-05 2.226427e-05
#>  [6,] 4.868770e-05  3.671503e-05 2.936036e-05 2.441896e-05 2.126214e-05
#>  [7,] 3.746395e-05  2.550115e-05 2.191653e-05 1.988535e-05 1.554748e-05
#>  [8,] 5.697002e-05  4.876151e-05 3.550205e-05 2.492008e-05 2.587264e-05
#>  [9,] 4.525675e-05  1.792573e-05 2.321763e-05 2.877659e-05 1.620055e-05
#> [10,] 1.584170e-05  2.644947e-05 1.336186e-05 3.698435e-06 1.124864e-05
#>                V55          V56           V57          V58          V59
#>  [1,] 1.988966e-05 2.064787e-05  2.689681e-05 1.554833e-05 1.854514e-05
#>  [2,] 1.906543e-05 1.892464e-05  2.469875e-05 1.421726e-05 1.708553e-05
#>  [3,] 1.150424e-05 1.049605e-05  2.007688e-05 4.825159e-06 7.523703e-06
#>  [4,] 2.299040e-05 2.429855e-05  4.921333e-05 7.823088e-06 1.469999e-05
#>  [5,] 1.946333e-05 2.147334e-05  2.718722e-05 1.558796e-05 1.932337e-05
#>  [6,] 1.727693e-05 1.847228e-05  1.978524e-05 1.591922e-05 1.812717e-05
#>  [7,] 1.381629e-05 1.450576e-05  1.697955e-05 1.160456e-05 1.362398e-05
#>  [8,] 1.935622e-05 2.084053e-05  1.776792e-05 1.996904e-05 2.146727e-05
#>  [9,] 1.954076e-05 1.994111e-05  3.216630e-05 1.157038e-05 1.587776e-05
#> [10,] 3.534398e-06 4.775740e-06 -1.301805e-06 8.304111e-06 7.409630e-06
#>                 V6          V60           V7           V8            V9
#>  [1,] 1.588537e-04 1.580751e-05 2.258840e-04 2.498883e-04  2.595518e-04
#>  [2,] 1.421032e-04 1.446798e-05 2.017709e-04 2.248915e-04  2.376660e-04
#>  [3,] 4.038664e-05 7.060833e-06 1.030819e-04 7.058177e-05 -1.060604e-04
#>  [4,] 6.177998e-05 1.683142e-05 2.634246e-04 1.284393e-04 -3.968764e-04
#>  [5,] 1.673954e-04 1.698780e-05 2.452182e-04 2.625465e-04  2.589123e-04
#>  [6,] 1.829076e-04 1.454346e-05 2.223198e-04 2.646833e-04  3.687218e-04
#>  [7,] 1.377590e-04 1.130228e-05 1.798225e-04 1.946663e-04  2.221432e-04
#>  [8,] 2.107745e-04 1.701032e-05 2.253904e-04 3.199211e-04  5.845950e-04
#>  [9,] 1.132874e-04 1.471228e-05 2.118751e-04 1.921074e-04  2.811765e-05
#> [10,] 1.183257e-04 4.283701e-06 7.659808e-05 1.545128e-04  3.893923e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,]  0.05870964  0.05574825  0.06302564  0.04945253  0.05090959  0.0529304
#> [2,] -0.05870964 -0.05574825 -0.06302564 -0.04945253 -0.05090959 -0.0529304
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.04305272  0.05745508  0.05779869  0.04082499
#> [2,] -0.04305272 -0.05745508 -0.05779869 -0.04082499
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5566751 0.6269366 0.7374124 0.3732215 0.4205510 0.3726686 0.1641842
#>  [8] 0.5840431 0.4724074 0.2583628
#> 
#> $B[[2]]
#> [1] -0.0121558 -0.1036993
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  2.126108e-03  1.989584e-03  1.426273e-03  3.867859e-03  2.285890e-03
#>  [6]  1.816864e-03  1.606741e-03  1.546200e-03  2.397968e-03 -3.640312e-05
#> 
#> $vB[[2]]
#> [1]  0.06509454 -0.06509454
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#>  [2,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [3,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#>  [4,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [5,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#>  [6,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#>  [7,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#>  [8,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#>  [9,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [10,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [11,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [12,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [13,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [14,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [15,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [16,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [17,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [18,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [19,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [20,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [21,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [22,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [23,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [24,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [25,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [26,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [27,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [28,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [29,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [30,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [31,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [32,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [33,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [34,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [35,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [36,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [37,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [38,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [39,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [40,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#>  [2,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [3,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#>  [4,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [5,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#>  [6,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#>  [7,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#>  [8,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#>  [9,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [10,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [11,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [12,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [13,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [14,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [15,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [16,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [17,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [18,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [19,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [20,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [21,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [22,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [23,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [24,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [25,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [26,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [27,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [28,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [29,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [30,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [31,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [32,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [33,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [34,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [35,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [36,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [37,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [38,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [39,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [40,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#>  [2,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [3,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#>  [4,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [5,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#>  [6,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#>  [7,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#>  [8,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#>  [9,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [10,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [11,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [12,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [13,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [14,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [15,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [16,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [17,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [18,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [19,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [20,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [21,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [22,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [23,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [24,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [25,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [26,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [27,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [28,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [29,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [30,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [31,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [32,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [33,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [34,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [35,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [36,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [37,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [38,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [39,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [40,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#>  [2,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [3,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#>  [4,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [5,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#>  [6,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#>  [7,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#>  [8,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#>  [9,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [10,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [11,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [12,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [13,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [14,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [15,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [16,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [17,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [18,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [19,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [20,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [21,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [22,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [23,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [24,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [25,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [26,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [27,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [28,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [29,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [30,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [31,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [32,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [33,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [34,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [35,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [36,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [37,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [38,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [39,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [40,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#>  [2,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [3,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#>  [4,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [5,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#>  [6,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#>  [7,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#>  [8,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#>  [9,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [10,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [11,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [12,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [13,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [14,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [15,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [16,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [17,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [18,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [19,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [20,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [21,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [22,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [23,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [24,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [25,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [26,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [27,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [28,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [29,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [30,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [31,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [32,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [33,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [34,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [35,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [36,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [37,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [38,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [39,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [40,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#>  [2,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [3,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#>  [4,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [5,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#>  [6,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#>  [7,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#>  [8,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#>  [9,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [10,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [11,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [12,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [13,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [14,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [15,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [16,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [17,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [18,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [19,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [20,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [21,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [22,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [23,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [24,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [25,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [26,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [27,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [28,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [29,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [30,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [31,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [32,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [33,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [34,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [35,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [36,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [37,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [38,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [39,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [40,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7565580 0.7972812 0.9067294 0.5578438 0.7537553 0.6047444 0.5438862
#>  [2,] 0.5225347 0.7233858 0.7818685 0.3591116 0.3528427 0.3314768 0.3224957
#>  [3,] 0.7988017 0.8434045 0.9263380 0.6331714 0.7600774 0.6422474 0.5384338
#>  [4,] 0.8634113 0.9103717 0.9605544 0.7242640 0.7187211 0.7354651 0.5374289
#>  [5,] 0.7617494 0.8079820 0.9099660 0.5791721 0.7600650 0.5993841 0.5173154
#>  [6,] 0.6519056 0.7690442 0.8791631 0.4225880 0.6719275 0.4815002 0.4418145
#>  [7,] 0.7198731 0.8144003 0.8741993 0.5282158 0.4355825 0.5728556 0.3771215
#>  [8,] 0.7817660 0.8692882 0.9128526 0.6162270 0.5044634 0.6121996 0.3883638
#>  [9,] 0.7555551 0.8184070 0.8895479 0.6493885 0.6934321 0.6166439 0.5071216
#> [10,] 0.7818338 0.8261551 0.9204083 0.6058624 0.7874812 0.6370480 0.5217061
#> [11,] 0.6113545 0.7524119 0.8576020 0.3737512 0.5956767 0.4320037 0.4021916
#> [12,] 0.7252705 0.8307481 0.8958926 0.4887738 0.5373635 0.4953953 0.4288239
#> [13,] 0.7704519 0.8533908 0.9039968 0.6106557 0.5940781 0.5766823 0.4830442
#> [14,] 0.8119974 0.8664356 0.9123403 0.6999441 0.6215773 0.6666836 0.5058137
#> [15,] 0.8570299 0.8889399 0.9471928 0.7196296 0.7257245 0.7191943 0.5429777
#> [16,] 0.8208703 0.8680844 0.9433105 0.6644651 0.7869925 0.6677091 0.5725404
#> [17,] 0.7822053 0.8625824 0.9167830 0.6049048 0.6923106 0.6232963 0.4971155
#> [18,] 0.7215411 0.8354485 0.8532655 0.5566769 0.3594337 0.5326285 0.3737685
#> [19,] 0.6630601 0.7269293 0.8267936 0.5296008 0.6015804 0.5117102 0.4518205
#> [20,] 0.7323031 0.8293794 0.8975098 0.5220495 0.6422878 0.5345245 0.4432123
#> [21,] 0.7551350 0.8458436 0.9031597 0.5874558 0.6710759 0.5625199 0.4587951
#> [22,] 0.8129820 0.8633376 0.9341851 0.6721824 0.7897040 0.6628275 0.5396042
#> [23,] 0.8114341 0.8803714 0.9360517 0.6789810 0.6042101 0.6444830 0.4371535
#> [24,] 0.7451905 0.8312805 0.9298813 0.4925563 0.6560186 0.5579906 0.4678661
#> [25,] 0.8086178 0.8720929 0.9366840 0.6683138 0.7644906 0.6509479 0.5038141
#> [26,] 0.7684081 0.8250175 0.9163496 0.5594509 0.6906049 0.6023958 0.5226868
#> [27,] 0.8316262 0.8787341 0.9492303 0.6737396 0.7877834 0.6718873 0.5703176
#> [28,] 0.7690527 0.8369165 0.9127718 0.5826987 0.7100312 0.6008393 0.5103347
#> [29,] 0.7717409 0.8206316 0.9155859 0.5691185 0.7302967 0.6077994 0.5270215
#> [30,] 0.7757896 0.8549665 0.9214978 0.6303235 0.6197931 0.6051521 0.4634554
#> [31,] 0.8033141 0.8409045 0.9233812 0.6044928 0.7034453 0.6623119 0.5030330
#> [32,] 0.7853305 0.8707853 0.9167184 0.6104771 0.6722347 0.6138757 0.4788949
#> [33,] 0.7541046 0.8029890 0.8985717 0.5788615 0.7319295 0.5836343 0.5048037
#> [34,] 0.8088989 0.8825408 0.9334691 0.6624463 0.6021512 0.6277802 0.4489646
#> [35,] 0.6231362 0.7543060 0.8620747 0.3922265 0.5753431 0.4416397 0.3923049
#> [36,] 0.6750894 0.8191619 0.8750568 0.4817535 0.4037242 0.4520549 0.3099987
#> [37,] 0.7532461 0.8639783 0.9147607 0.5887887 0.6273594 0.5599024 0.4486846
#> [38,] 0.7226360 0.8361900 0.8959328 0.5596412 0.6269711 0.5113958 0.4094695
#> [39,] 0.7768549 0.8639835 0.9189366 0.6415544 0.5268028 0.6127777 0.3860975
#> [40,] 0.7087663 0.8134507 0.9019720 0.5141129 0.5555430 0.5572138 0.3966846
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7893434 0.7947194 0.4304517
#>  [2,] 0.5085665 0.5899164 0.2182457
#>  [3,] 0.8327651 0.8349928 0.4415519
#>  [4,] 0.8785318 0.8893079 0.4256445
#>  [5,] 0.8038245 0.8138040 0.4310771
#>  [6,] 0.7096336 0.7296214 0.3071463
#>  [7,] 0.7274387 0.7153695 0.3382673
#>  [8,] 0.8032025 0.8079780 0.3321940
#>  [9,] 0.7789147 0.7918815 0.4314399
#> [10,] 0.8285437 0.8273605 0.4241368
#> [11,] 0.6537885 0.6817104 0.2802458
#> [12,] 0.7427082 0.7736027 0.2707506
#> [13,] 0.7870210 0.8015819 0.3625182
#> [14,] 0.8155898 0.8328787 0.4190903
#> [15,] 0.8729080 0.8773540 0.4429434
#> [16,] 0.8463919 0.8619735 0.4436924
#> [17,] 0.8123758 0.8189380 0.3413347
#> [18,] 0.6973700 0.7312107 0.3111945
#> [19,] 0.6745248 0.6856679 0.4247703
#> [20,] 0.7680860 0.7762086 0.3078041
#> [21,] 0.7961363 0.7993104 0.3350628
#> [22,] 0.8565406 0.8603146 0.4389666
#> [23,] 0.8395014 0.8415498 0.4225799
#> [24,] 0.7497206 0.7954811 0.3278983
#> [25,] 0.8553476 0.8638814 0.4138026
#> [26,] 0.7887013 0.8036029 0.4113118
#> [27,] 0.8543846 0.8702295 0.4317997
#> [28,] 0.8013337 0.8161655 0.3818643
#> [29,] 0.8070009 0.8081520 0.4075196
#> [30,] 0.7889147 0.8034827 0.4111192
#> [31,] 0.8338362 0.8376354 0.4057143
#> [32,] 0.8079716 0.8219365 0.3284714
#> [33,] 0.7912841 0.8011683 0.4262289
#> [34,] 0.8272041 0.8408550 0.3809900
#> [35,] 0.6577644 0.6830996 0.2842382
#> [36,] 0.6923704 0.7101232 0.2570392
#> [37,] 0.7812194 0.8147234 0.3053928
#> [38,] 0.7702634 0.7814673 0.3184065
#> [39,] 0.8133520 0.8142802 0.3236037
#> [40,] 0.7231439 0.7761653 0.2963018
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6685067 0.3314933
#>  [2,] 0.6421319 0.3578681
#>  [3,] 0.6742824 0.3257176
#>  [4,] 0.6820149 0.3179851
#>  [5,] 0.6700192 0.3299808
#>  [6,] 0.6584244 0.3415756
#>  [7,] 0.6602175 0.3397825
#>  [8,] 0.6699500 0.3300500
#>  [9,] 0.6704889 0.3295111
#> [10,] 0.6727986 0.3272014
#> [11,] 0.6527613 0.3472387
#> [12,] 0.6633113 0.3366887
#> [13,] 0.6699332 0.3300668
#> [14,] 0.6751377 0.3248623
#> [15,] 0.6805146 0.3194854
#> [16,] 0.6779225 0.3220775
#> [17,] 0.6725572 0.3274428
#> [18,] 0.6604340 0.3395660
#> [19,] 0.6567198 0.3432802
#> [20,] 0.6657713 0.3342287
#> [21,] 0.6696936 0.3303064
#> [22,] 0.6775248 0.3224752
#> [23,] 0.6750540 0.3249460
#> [24,] 0.6668276 0.3331724
#> [25,] 0.6773908 0.3226092
#> [26,] 0.6691036 0.3308964
#> [27,] 0.6791088 0.3208912
#> [28,] 0.6707557 0.3292443
#> [29,] 0.6700371 0.3299629
#> [30,] 0.6712502 0.3287498
#> [31,] 0.6728610 0.3271390
#> [32,] 0.6727690 0.3272310
#> [33,] 0.6686758 0.3313242
#> [34,] 0.6747482 0.3252518
#> [35,] 0.6535225 0.3464775
#> [36,] 0.6575547 0.3424453
#> [37,] 0.6705665 0.3294335
#> [38,] 0.6666056 0.3333944
#> [39,] 0.6713936 0.3286064
#> [40,] 0.6640418 0.3359582
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]     [,3]        [,4]       [,5]        [,6]
#>  [1,] 1.1339007 1.3693874 2.274339  0.23241582  1.1187423  0.42527317
#>  [2,] 0.0901999 0.9613187 1.276589 -0.57922196 -0.6065671 -0.70151321
#>  [3,] 1.3788214 1.6837807 2.531752  0.54584636  1.1531038  0.58513209
#>  [4,] 1.8439169 2.3181815 3.192589  0.96571232  0.9381267  1.02252967
#>  [5,] 1.1622945 1.4369506 2.313220  0.31937561  1.1530361  0.40289935
#>  [6,] 0.6274260 1.2029220 1.984529 -0.31215830  0.7169157 -0.07403313
#>  [7,] 0.9438323 1.4788601 1.938609  0.11298333 -0.2591100  0.29351138
#>  [8,] 1.2759873 1.8946800 2.348974  0.47356428  0.0178540  0.45656761
#>  [9,] 1.1284628 1.5055917 2.086131  0.61635244  0.8162142  0.47532725
#> [10,] 1.2763852 1.5586188 2.447907  0.42995271  1.3098088  0.56257438
#> [11,] 0.4530096 1.1115174 1.795514 -0.51615749  0.3874831 -0.27368082
#> [12,] 0.9707578 1.5909386 2.152398 -0.04491228  0.1497330 -0.01841935
#> [13,] 1.2108649 1.7614471 2.242444  0.45006939  0.3808501  0.30916870
#> [14,] 1.4630415 1.8698036 2.342550  0.84703157  0.4962483  0.69322338
#> [15,] 1.7908374 2.0799576 2.886855  0.94262501  0.9730372  0.94046862
#> [16,] 1.5222548 1.8841268 2.811807  0.68325633  1.3068915  0.69784184
#> [17,] 1.2785641 1.8369066 2.399418  0.42594438  0.8109439  0.50356299
#> [18,] 0.9521189 1.6247448 1.760446  0.22768622 -0.5778229  0.13069986
#> [19,] 0.6769612 0.9790987 1.563071  0.11854179  0.4120544  0.04684945
#> [20,] 1.0063392 1.5812353 2.169857  0.08825538  0.5853076  0.13831814
#> [21,] 1.1261894 1.7023670 2.232836  0.35345780  0.7130552  0.25139519
#> [22,] 1.4695039 1.8432922 2.652829  0.71807221  1.3231422  0.67592005
#> [23,] 1.4593556 1.9959518 2.683596  0.74909266  0.4230384  0.59487524
#> [24,] 1.0731238 1.5947298 2.584868 -0.02977683  0.6456013  0.23301099
#> [25,] 1.4410539 1.9195918 2.694208  0.70056851  1.1774589  0.62320823
#> [26,] 1.1993439 1.5507183 2.393752  0.23893397  0.8029488  0.41545757
#> [27,] 1.5971967 1.9804970 2.928351  0.72514792  1.3116154  0.71673327
#> [28,] 1.2029699 1.6354621 2.347958  0.33386167  0.8955357  0.40896332
#> [29,] 1.2181673 1.5206322 2.383829  0.27825566  0.9961284  0.43807150
#> [30,] 1.2412967 1.7740972 2.462874  0.53360505  0.4886699  0.42697936
#> [31,] 1.4071378 1.6649733 2.489200  0.42422080  0.8637587  0.67361413
#> [32,] 1.2970049 1.9079199 2.398572  0.44931799  0.7183097  0.46363314
#> [33,] 1.1206250 1.4050812 2.181454  0.31810150  1.0044341  0.33771067
#> [34,] 1.4428712 2.0167139 2.641241  0.67421509  0.4144364  0.52270542
#> [35,] 0.5028811 1.1217111 1.832629 -0.43796264  0.3036848 -0.23450992
#> [36,] 0.7312950 1.5106795 1.946430 -0.07301859 -0.3899713 -0.19237152
#> [37,] 1.1160006 1.8487333 2.373200  0.35896034  0.5209052  0.24076593
#> [38,] 0.9575748 1.6301487 2.152829  0.23970585  0.5192444  0.04559110
#> [39,] 1.2474314 1.8487776 2.427986  0.58211727  0.1073141  0.45900348
#> [40,] 0.8893998 1.4725901 2.219331  0.05646674  0.2230925  0.22986184
#>              [,7]       [,8]      [,9]      [,10]
#>  [1,]  0.17599771 1.32097202 1.3536111 -0.2800083
#>  [2,] -0.74232589 0.03426927 0.3636199 -1.2759193
#>  [3,]  0.15403918 1.60535188 1.6214339 -0.2348660
#>  [4,]  0.14999625 1.97859996 2.0836914 -0.2996439
#>  [5,]  0.06928935 1.41037161 1.4749198 -0.2774578
#>  [6,] -0.23380103 0.89360517 0.9927027 -0.8134948
#>  [7,] -0.50178393 0.98166643 0.9216073 -0.6710255
#>  [8,] -0.45419525 1.40643145 1.4369247 -0.6982786
#>  [9,]  0.02848813 1.25935281 1.3363042 -0.2759788
#> [10,]  0.08687913 1.57534141 1.5670351 -0.3058140
#> [11,] -0.39634164 0.63573398 0.7616432 -0.9432426
#> [12,] -0.28665106 1.06009246 1.2287669 -0.9908175
#> [13,] -0.06784916 1.30706150 1.3962108 -0.5644510
#> [14,]  0.02325594 1.48674902 1.6061682 -0.3265091
#> [15,]  0.17233605 1.92691849 1.9676081 -0.2292249
#> [16,]  0.29222369 1.70657788 1.8317792 -0.2261897
#> [17,] -0.01153822 1.46552233 1.5091690 -0.6573520
#> [18,] -0.51608360 0.83480518 1.0007741 -0.7945410
#> [19,] -0.19331763 0.72872229 0.7799432 -0.3032209
#> [20,] -0.22813521 1.19753508 1.2437067 -0.8104055
#> [21,] -0.16519442 1.36231894 1.3819898 -0.6853746
#> [22,]  0.15874941 1.78684999 1.8179053 -0.2453570
#> [23,] -0.25272263 1.65452306 1.6698052 -0.3121917
#> [24,] -0.12871301 1.09712246 1.3582865 -0.7177063
#> [25,]  0.01525666 1.77717451 1.8479089 -0.3482675
#> [26,]  0.09080944 1.31711504 1.4089665 -0.3585449
#> [27,]  0.28314690 1.76941222 1.9029897 -0.2745123
#> [28,]  0.04134460 1.39465089 1.4905811 -0.4816428
#> [29,]  0.10819153 1.43063907 1.4380467 -0.3742286
#> [30,] -0.14643972 1.31839572 1.4082052 -0.3593405
#> [31,]  0.01213206 1.61306303 1.6407382 -0.3817110
#> [32,] -0.08447058 1.43688352 1.5295227 -0.7151069
#> [33,]  0.01921557 1.33268292 1.3936121 -0.2972542
#> [34,] -0.20485516 1.56594057 1.6646037 -0.4853485
#> [35,] -0.43763372 0.65334723 0.7680534 -0.9235348
#> [36,] -0.80012519 0.81122448 0.8959824 -1.0614148
#> [37,] -0.20598701 1.27278681 1.4809986 -0.8217478
#> [38,] -0.36615882 1.20979916 1.2742378 -0.7611047
#> [39,] -0.46374615 1.47193970 1.4780657 -0.7372596
#> [40,] -0.41929862 0.96011033 1.2434578 -0.8649712
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1984998 -0.5029392
#>  [2,] 0.1583248 -0.4263046
#>  [3,] 0.2013175 -0.5263009
#>  [4,] 0.2001061 -0.5629410
#>  [5,] 0.1981158 -0.5101561
#>  [6,] 0.1840063 -0.4722743
#>  [7,] 0.1691943 -0.4950695
#>  [8,] 0.1764101 -0.5315488
#>  [9,] 0.1942634 -0.5161336
#> [10,] 0.1991997 -0.5216703
#> [11,] 0.1768883 -0.4543105
#> [12,] 0.1801236 -0.4979621
#> [13,] 0.1894215 -0.5184613
#> [14,] 0.1921289 -0.5393866
#> [15,] 0.2006521 -0.5554857
#> [16,] 0.2062724 -0.5379686
#> [17,] 0.1925064 -0.5272675
#> [18,] 0.1676266 -0.4976025
#> [19,] 0.1822314 -0.4664790
#> [20,] 0.1858673 -0.5032533
#> [21,] 0.1899435 -0.5168560
#> [22,] 0.2034380 -0.5389819
#> [23,] 0.1888671 -0.5422668
#> [24,] 0.1887979 -0.5050734
#> [25,] 0.1997530 -0.5420538
#> [26,] 0.1955036 -0.5086300
#> [27,] 0.2064300 -0.5432494
#> [28,] 0.1951538 -0.5164510
#> [29,] 0.1969452 -0.5114076
#> [30,] 0.1906085 -0.5232367
#> [31,] 0.1929134 -0.5282402
#> [32,] 0.1908549 -0.5298809
#> [33,] 0.1959111 -0.5062907
#> [34,] 0.1886951 -0.5410448
#> [35,] 0.1754007 -0.4591582
#> [36,] 0.1654284 -0.4869874
#> [37,] 0.1874998 -0.5232484
#> [38,] 0.1850276 -0.5078450
#> [39,] 0.1767815 -0.5377135
#> [40,] 0.1755168 -0.5058417
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.6685067  0.6685067
#>  [2,]  0.3578681 -0.3578681
#>  [3,] -0.6742824  0.6742824
#>  [4,] -0.6820149  0.6820149
#>  [5,]  0.3299808 -0.3299808
#>  [6,]  0.3415756 -0.3415756
#>  [7,] -0.6602175  0.6602175
#>  [8,]  0.3300500 -0.3300500
#>  [9,] -0.6704889  0.6704889
#> [10,]  0.3272014 -0.3272014
#> [11,]  0.3472387 -0.3472387
#> [12,]  0.3366887 -0.3366887
#> [13,] -0.6699332  0.6699332
#> [14,] -0.6751377  0.6751377
#> [15,] -0.6805146  0.6805146
#> [16,] -0.6779225  0.6779225
#> [17,]  0.3274428 -0.3274428
#> [18,]  0.3395660 -0.3395660
#> [19,] -0.6567198  0.6567198
#> [20,]  0.3342287 -0.3342287
#> [21,]  0.3303064 -0.3303064
#> [22,]  0.3224752 -0.3224752
#> [23,] -0.6750540  0.6750540
#> [24,] -0.6668276  0.6668276
#> [25,]  0.3226092 -0.3226092
#> [26,] -0.6691036  0.6691036
#> [27,] -0.6791088  0.6791088
#> [28,]  0.3292443 -0.3292443
#> [29,] -0.6700371  0.6700371
#> [30,]  0.3287498 -0.3287498
#> [31,] -0.6728610  0.6728610
#> [32,]  0.3272310 -0.3272310
#> [33,] -0.6686758  0.6686758
#> [34,]  0.3252518 -0.3252518
#> [35,]  0.3464775 -0.3464775
#> [36,] -0.6575547  0.6575547
#> [37,]  0.3294335 -0.3294335
#> [38,]  0.3333944 -0.3333944
#> [39,]  0.3286064 -0.3286064
#> [40,] -0.6640418  0.6640418
#> 
#> $L
#> [1] 0.7193957 0.6768985 0.7149891 0.6936390 0.7236611 0.7401569
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319