Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12           V13
#>  [1,] -0.1972335 -0.11163579 -0.021551177 -0.02093148  0.0635198735
#>  [2,] -0.3345897 -0.07229727 -0.089954379 -0.03736870 -0.0438055547
#>  [3,] -0.2463914 -0.03545037 -0.082768903 -0.16188605 -0.0438137847
#>  [4,] -0.1911544 -0.23663772 -0.107708678 -0.01989743 -0.1661329120
#>  [5,] -0.4146850 -0.17521868 -0.115476698 -0.07657037 -0.1504767865
#>  [6,] -0.2523797 -0.04595920 -0.050803325 -0.16605515 -0.1793441131
#>  [7,] -0.3641645 -0.06994914 -0.039450317 -0.05559301 -0.0031861106
#>  [8,] -0.4277082 -0.05813883 -0.064010470 -0.06602036 -0.1226291639
#>  [9,] -0.3794586 -0.02680216 -0.143737440 -0.08126050  0.0077383452
#> [10,] -0.2358254 -0.15385491 -0.007976543 -0.16449584 -0.0005653491
#>                V14          V15         V16          V17         V18
#>  [1,]  0.009727559 -0.037269063  0.06475943 -0.021067819  0.05161383
#>  [2,] -0.092612583 -0.002079786 -0.05252456 -0.073383843 -0.01693957
#>  [3,] -0.175118174 -0.115476436 -0.13427982 -0.034769508 -0.00638224
#>  [4,] -0.003938653 -0.123035191 -0.13245547  0.060033964  0.09556470
#>  [5,] -0.087713462 -0.016661790 -0.11859777 -0.019579633 -0.03304712
#>  [6,] -0.137592465 -0.054680472 -0.04517771 -0.004068404  0.07410452
#>  [7,] -0.087151313 -0.023727243  0.02110826  0.029963935 -0.08444593
#>  [8,]  0.021256097 -0.049121185 -0.02858558  0.083039795  0.02243551
#>  [9,]  0.023515041 -0.003495788  0.06935395  0.052763600 -0.02613254
#> [10,] -0.055590320  0.062716733  0.02802204  0.052532490  0.06567619
#>               V19         V2        V20        V21         V22          V23
#>  [1,]  0.04599707 -0.2341554 0.08380145 0.13294866  0.17253651  0.133870635
#>  [2,]  0.02222739 -0.3153787 0.13324481 0.06921673  0.11529191  0.097343897
#>  [3,] -0.01374919 -0.1475578 0.03878394 0.05325251 -0.02286846  0.045705580
#>  [4,]  0.07373643 -0.2051726 0.13075862 0.05987017  0.02266729  0.024270011
#>  [5,] -0.06775439 -0.3980632 0.01861422 0.17669378  0.19489507  0.209253592
#>  [6,]  0.07648224 -0.3570672 0.05177745 0.12258670  0.14397922  0.101420305
#>  [7,]  0.12195152 -0.3186520 0.11946795 0.07880001 -0.02944267 -0.002479346
#>  [8,]  0.10960213 -0.2949132 0.12210809 0.21380328  0.25651259  0.184789153
#>  [9,]  0.14211062 -0.3713391 0.11888176 0.16057823  0.11725629  0.143265633
#> [10,]  0.07596669 -0.1423331 0.15284310 0.01843023  0.03543018  0.180485497
#>              V24        V25         V26         V27        V28         V29
#>  [1,] 0.15852044 0.11391648  0.08236440  0.19505151 0.12281357  0.03603780
#>  [2,] 0.13067724 0.22578263  0.28651244  0.27775526 0.17603406  0.20246169
#>  [3,] 0.01769971 0.11214639 -0.02848380 -0.07983915 0.01954712 -0.05046120
#>  [4,] 0.19563284 0.09806485  0.03575093  0.14899478 0.03745917  0.09918252
#>  [5,] 0.14169329 0.17109515  0.13591731  0.19678747 0.17205165  0.11569356
#>  [6,] 0.21609628 0.17460733  0.13305740  0.08131374 0.13748425  0.04489342
#>  [7,] 0.06622934 0.09740523  0.08022466  0.16874701 0.05595072  0.12456307
#>  [8,] 0.10847704 0.21074664  0.24631610  0.26526531 0.28071131  0.19891683
#>  [9,] 0.12388097 0.16891221  0.19480867  0.23385547 0.26826181  0.21710231
#> [10,] 0.09639424 0.06814092  0.11950389  0.17348888 0.11223963 -0.01438508
#>               V3         V30          V31          V32          V33         V34
#>  [1,] -0.2362815  0.06842592  0.049997050  0.007187536  0.098848641  0.12313052
#>  [2,] -0.3138384  0.16265718  0.136858546  0.105408103  0.085610380  0.07530118
#>  [3,] -0.3394421 -0.14082914 -0.154968411 -0.109391331 -0.104649159 -0.04053461
#>  [4,] -0.3013142  0.09251760 -0.063533767 -0.137332460 -0.015466953 -0.10001436
#>  [5,] -0.2789006  0.05960093 -0.069579658 -0.041266048 -0.150413187 -0.01154807
#>  [6,] -0.3720544 -0.03436881 -0.071071703 -0.051151798 -0.120864846 -0.10968743
#>  [7,] -0.2078708  0.07675426 -0.091612865 -0.059424957 -0.027804874 -0.08115759
#>  [8,] -0.3657913  0.02817849  0.058456596 -0.007142749 -0.001457825 -0.06259816
#>  [9,] -0.2081172  0.06013181  0.044417338  0.119365040  0.040478594  0.01119471
#> [10,] -0.1469968 -0.06577102  0.001397246 -0.028914136 -0.064673278 -0.05016540
#>                V35          V36          V37          V38          V39
#>  [1,] -0.034511020 -0.025161823  0.089090655  0.003743283  0.069840311
#>  [2,]  0.007071982  0.008120127 -0.012208583 -0.044631215 -0.143484398
#>  [3,] -0.173730965  0.025252990 -0.023985656 -0.163500578 -0.045300478
#>  [4,]  0.032211681 -0.150441990 -0.001027065 -0.132984234 -0.198971222
#>  [5,] -0.070068353 -0.176854651 -0.132157710 -0.102331152 -0.069726551
#>  [6,] -0.018711089 -0.012656434 -0.113982767 -0.100891945 -0.122734967
#>  [7,] -0.041776802 -0.084508394 -0.104288622 -0.058017397 -0.091886564
#>  [8,]  0.087760323  0.069345691 -0.036887566  0.012585921  0.008024456
#>  [9,] -0.006435431  0.097268901  0.139254993  0.016307216  0.012670512
#> [10,] -0.088551017 -0.081670484  0.003360163 -0.038154533 -0.141161703
#>               V4          V40           V41          V42         V43
#>  [1,] -0.2274240 -0.097001433 -0.0134915571 -0.033523423 -0.13946500
#>  [2,] -0.2626151 -0.072079176 -0.1355957826 -0.173198880 -0.08629977
#>  [3,] -0.2373591 -0.152852888 -0.2047899325 -0.044991263 -0.05094451
#>  [4,] -0.3086450 -0.094175876 -0.0678670102 -0.109192125 -0.12766670
#>  [5,] -0.3668231 -0.149730197 -0.0511635493 -0.054947143 -0.16114752
#>  [6,] -0.2639144 -0.051415047 -0.0703326383 -0.108201953 -0.10130983
#>  [7,] -0.1753754 -0.073622089 -0.1731727002 -0.126553509 -0.03067690
#>  [8,] -0.3067810  0.002323037 -0.1445874912  0.002524237 -0.13947358
#>  [9,] -0.3447093  0.006134027  0.0006101594  0.018008994  0.02789796
#> [10,] -0.1518873 -0.061992115 -0.0302355092 -0.167266066 -0.15038477
#>               V44         V45         V46         V47         V48        V49
#>  [1,] -0.04744655 -0.06456974 -0.17190250 -0.19275867 -0.21219558 -0.2141837
#>  [2,] -0.15848034 -0.01889578 -0.16428562 -0.11604472 -0.21799421 -0.1529173
#>  [3,] -0.10301013 -0.14228342 -0.04271420 -0.26955359 -0.14668597 -0.2243459
#>  [4,] -0.06908265 -0.10259796 -0.07814167 -0.12214724 -0.18513531 -0.1493496
#>  [5,] -0.02913182 -0.13037703 -0.20463944 -0.16763141 -0.30434163 -0.3669053
#>  [6,] -0.12283111 -0.07801746 -0.14627837 -0.26773432 -0.28279018 -0.3375650
#>  [7,] -0.20135793 -0.08084388 -0.06269386 -0.27111401 -0.12222023 -0.2536493
#>  [8,] -0.07429175 -0.13981695 -0.17812254 -0.15792450 -0.32317121 -0.3433650
#>  [9,] -0.05242690 -0.13812823 -0.17552120 -0.09963786 -0.26969106 -0.3430717
#> [10,] -0.14512134 -0.14985576 -0.11308401 -0.20830620 -0.08479271 -0.3117771
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.2144845 -0.4156410 -0.2814103 -0.3922507 -0.3195951 -0.3204601
#>  [2,] -0.2603682 -0.2088715 -0.2822618 -0.1953443 -0.3225768 -0.3009202
#>  [3,] -0.2285483 -0.2780229 -0.2263627 -0.2357617 -0.3383698 -0.2668784
#>  [4,] -0.2857329 -0.2722938 -0.3321487 -0.3498116 -0.3489026 -0.2384108
#>  [5,] -0.2334057 -0.3045222 -0.3748606 -0.3856351 -0.4115420 -0.4350228
#>  [6,] -0.2029410 -0.3438724 -0.3133274 -0.3525941 -0.3404741 -0.3033849
#>  [7,] -0.1603859 -0.3983294 -0.2549481 -0.3718004 -0.2537417 -0.3134520
#>  [8,] -0.3166847 -0.3875870 -0.4373136 -0.4014134 -0.4601105 -0.3594382
#>  [9,] -0.1454511 -0.3902138 -0.3658708 -0.3969601 -0.3468334 -0.4312299
#> [10,] -0.2467178 -0.1922921 -0.2985601 -0.3135431 -0.1674661 -0.3302964
#>              V55        V56        V57        V58        V59         V6
#>  [1,] -0.3810538 -0.3312418 -0.4172255 -0.4202690 -0.2869069 -0.1421945
#>  [2,] -0.3434220 -0.2678879 -0.2456669 -0.3521957 -0.3151181 -0.0804283
#>  [3,] -0.2111944 -0.2221026 -0.2276837 -0.2233654 -0.3097800 -0.2464584
#>  [4,] -0.3722182 -0.3773699 -0.3296005 -0.2898276 -0.2823476 -0.1088669
#>  [5,] -0.3892036 -0.3444077 -0.4556457 -0.2836162 -0.4588065 -0.1919683
#>  [6,] -0.2080163 -0.3617557 -0.2447789 -0.3739474 -0.2722434 -0.1341461
#>  [7,] -0.2695452 -0.2762922 -0.3318684 -0.2887250 -0.4270960 -0.2162560
#>  [8,] -0.4728501 -0.3689311 -0.3531883 -0.4849131 -0.3448739 -0.2935090
#>  [9,] -0.3257815 -0.4666667 -0.3910285 -0.2913686 -0.2926629 -0.1917212
#> [10,] -0.2854220 -0.3739468 -0.2886659 -0.2982118 -0.2705078 -0.2419314
#>              V60          V7          V8          V9
#>  [1,] -0.2992810 -0.19585811 -0.17976416 -0.16426167
#>  [2,] -0.3164939 -0.11907475 -0.15815354 -0.08164409
#>  [3,] -0.3406322 -0.18374088 -0.19891811 -0.19742049
#>  [4,] -0.3534116 -0.07000323 -0.11516298 -0.20205712
#>  [5,] -0.3079474 -0.25220086 -0.26474370 -0.17634632
#>  [6,] -0.2558418 -0.18650856 -0.19981108 -0.15761180
#>  [7,] -0.4230014 -0.24398727 -0.19263525 -0.02799604
#>  [8,] -0.4546278 -0.08071879 -0.21046610 -0.13871875
#>  [9,] -0.3809996 -0.21829486 -0.23015905 -0.03612893
#> [10,] -0.2634474 -0.19487808 -0.07904417 -0.17027622
#> 
#> $W[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,]  0.09862801 -0.07345913 0.06251364 -0.10316494  0.08401982 -0.05775243
#> [2,] -0.06025300 -0.06757850 0.03928394 -0.08476298 -0.01476869  0.02959837
#>              [,7]        [,8]        [,9]       [,10]
#> [1,] -0.035418779 -0.06310531  0.11824980  0.04818164
#> [2,] -0.003989319  0.04413798 -0.01300987 -0.01584745
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.730881e-04 -0.0007578155 -0.0011036895 -0.0010416092 -0.0009661165
#>  [2,]  1.700769e-05  0.0001386592  0.0001489285  0.0001761251  0.0001523478
#>  [3,] -5.010468e-05 -0.0002449190 -0.0004016633 -0.0004049654 -0.0003239060
#>  [4,]  2.143962e-05  0.0001531973  0.0001787949  0.0002147190  0.0002017787
#>  [5,] -1.466691e-04 -0.0006478208 -0.0009914569 -0.0009442707 -0.0008672198
#>  [6,]  1.297434e-04  0.0006987974  0.0009654211  0.0009978406  0.0009053454
#>  [7,]  5.967621e-05  0.0003535169  0.0004623895  0.0005010619  0.0004532665
#>  [8,]  1.398766e-04  0.0006739297  0.0008881683  0.0008713596  0.0007987939
#>  [9,] -1.210299e-04 -0.0004697395 -0.0007069517 -0.0006598723 -0.0006247152
#> [10,] -9.951920e-05 -0.0004537372 -0.0007018627 -0.0006611243 -0.0005891884
#>                 V14           V15           V16           V17           V18
#>  [1,] -0.0007624917 -2.981071e-04 -0.0001845363 -1.376101e-05 -5.992436e-05
#>  [2,]  0.0001631455  2.287475e-04  0.0002654033  3.059376e-04  3.210067e-04
#>  [3,] -0.0001763138  6.140252e-05  0.0001571996  2.282013e-04  1.578980e-04
#>  [4,]  0.0002060620  2.596901e-04  0.0003098334  3.612016e-04  3.717825e-04
#>  [5,] -0.0006058039 -7.339915e-05  0.0001033573  3.122616e-04  2.287270e-04
#>  [6,]  0.0007264266  4.726102e-04  0.0004388508  3.950104e-04  4.796618e-04
#>  [7,]  0.0003977187  3.533795e-04  0.0003783401  4.019887e-04  4.426235e-04
#>  [8,]  0.0007309060  5.771722e-04  0.0005819037  5.419285e-04  5.376432e-04
#>  [9,] -0.0005165940 -2.304776e-04 -0.0001709688 -6.966582e-05 -7.419045e-05
#> [10,] -0.0003797701  1.862935e-05  0.0001516990  3.089782e-04  2.354826e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -4.344338e-04 -2.255460e-04 -4.713784e-04 -0.0005132712 -2.478282e-04
#>  [2,]  3.843254e-04  1.185967e-05  5.504814e-04  0.0005951267  5.473825e-04
#>  [3,]  5.461568e-06 -6.286673e-05 -6.604383e-06 -0.0000382619  4.801882e-05
#>  [4,]  4.409175e-04  2.003321e-05  5.535351e-04  0.0006009944  5.917809e-04
#>  [5,] -1.557183e-04 -1.852816e-04 -1.656134e-04 -0.0002135144  1.647394e-05
#>  [6,]  8.518377e-04  1.518005e-04  1.079864e-03  0.0011992944  1.010192e-03
#>  [7,]  6.205647e-04  6.540187e-05  8.016149e-04  0.0008828536  8.073223e-04
#>  [8,]  8.190231e-04  1.666215e-04  1.093163e-03  0.0011769203  9.364630e-04
#>  [9,] -2.814930e-04 -1.611230e-04 -2.456690e-04 -0.0002532777 -8.309663e-05
#> [10,] -3.710119e-05 -1.291290e-04 -2.445883e-05 -0.0000494282  1.349582e-04
#>                 V23           V24          V25          V26          V27
#>  [1,] -3.612565e-04 -2.728966e-04 0.0003026352 0.0005759196 0.0001122234
#>  [2,]  4.202629e-04  4.067774e-04 0.0004549138 0.0004658493 0.0004549624
#>  [3,]  7.107378e-05  1.690589e-04 0.0004192486 0.0005930300 0.0004163980
#>  [4,]  5.169071e-04  5.131890e-04 0.0005417465 0.0005581974 0.0005781873
#>  [5,] -1.317197e-04 -6.366150e-06 0.0006134474 0.0009185051 0.0004708834
#>  [6,]  9.319197e-04  8.382030e-04 0.0005040463 0.0003344015 0.0006527376
#>  [7,]  7.027174e-04  6.678806e-04 0.0005834168 0.0005312810 0.0006447820
#>  [8,]  8.077408e-04  7.541570e-04 0.0005131409 0.0003803569 0.0006340244
#>  [9,] -1.772959e-04 -1.252715e-04 0.0002551261 0.0004466652 0.0001404958
#> [10,]  4.059040e-05  1.344563e-04 0.0005850277 0.0008151471 0.0005018725
#>                 V28          V29            V3          V30          V31
#>  [1,] -3.526857e-04 0.0001073650 -2.712070e-04 0.0003570926 6.018136e-04
#>  [2,]  4.706650e-04 0.0005052138  2.490931e-06 0.0005023375 3.575026e-04
#>  [3,]  1.853661e-04 0.0003140658 -6.290975e-05 0.0004216362 5.754108e-04
#>  [4,]  5.948841e-04 0.0005902698  1.900557e-05 0.0005498642 4.081169e-04
#>  [5,]  7.304672e-06 0.0004531518 -2.221201e-04 0.0007287482 9.517949e-04
#>  [6,]  9.886143e-04 0.0007046707  1.663697e-04 0.0004685228 6.210041e-05
#>  [7,]  7.779455e-04 0.0006964156  6.681486e-05 0.0005937450 3.308425e-04
#>  [8,]  9.155603e-04 0.0006897158  1.881923e-04 0.0005956714 2.767030e-04
#>  [9,] -1.610158e-04 0.0001601042 -2.001220e-04 0.0002962468 4.033003e-04
#> [10,]  1.726350e-04 0.0004990327 -1.518189e-04 0.0006918418 8.325097e-04
#>                V32           V33           V34           V35           V36
#>  [1,] 0.0004504127  7.370182e-04  1.251700e-03  1.369014e-03  1.405915e-03
#>  [2,] 0.0003276106  2.814737e-04  3.105155e-04  3.123276e-04  2.437239e-04
#>  [3,] 0.0004820765  5.759027e-04  7.913442e-04  8.409376e-04  8.073778e-04
#>  [4,] 0.0003573602  2.938683e-04  2.892380e-04  2.819569e-04  2.231684e-04
#>  [5,] 0.0007445626  9.931749e-04  1.516674e-03  1.642450e-03  1.646108e-03
#>  [6,] 0.0001127096 -1.622907e-04 -5.051114e-04 -5.960317e-04 -6.876696e-04
#>  [7,] 0.0003086245  1.703993e-04  5.768792e-05  2.345464e-05 -4.823491e-05
#>  [8,] 0.0003247130  7.171875e-05 -1.859307e-04 -2.468066e-04 -4.081106e-04
#>  [9,] 0.0002933004  4.626695e-04  7.974662e-04  8.637304e-04  9.035498e-04
#> [10,] 0.0006722814  8.452211e-04  1.231832e-03  1.329161e-03  1.293732e-03
#>                 V37          V38          V39            V4          V40
#>  [1,]  0.0012025056 0.0002899445 2.054338e-06 -3.403484e-04 0.0001611059
#>  [2,]  0.0002833552 0.0002734059 2.434268e-04  1.117104e-05 0.0001633440
#>  [3,]  0.0006573322 0.0002503371 1.099535e-04 -9.000423e-05 0.0001935516
#>  [4,]  0.0002677531 0.0002699802 2.488571e-04  2.890737e-05 0.0001931944
#>  [5,]  0.0014288283 0.0004659677 1.414977e-04 -2.801033e-04 0.0002955023
#>  [6,] -0.0004447513 0.0001736391 3.516090e-04  2.216007e-04 0.0001438460
#>  [7,]  0.0000727493 0.0002821891 3.196654e-04  9.387826e-05 0.0001913241
#>  [8,] -0.0002579377 0.0002629839 3.556755e-04  2.412004e-04 0.0001194266
#>  [9,]  0.0008073580 0.0002119463 2.788540e-05 -2.456785e-04 0.0001265266
#> [10,]  0.0011118399 0.0004225254 1.714461e-04 -1.916457e-04 0.0002629007
#>                 V41           V42           V43           V44           V45
#>  [1,] -4.028374e-04 -8.169953e-04 -0.0007853173 -0.0008513700 -0.0009439417
#>  [2,]  1.084272e-04  8.570262e-05  0.0001219024  0.0001198820  0.0001808684
#>  [3,] -8.326509e-06 -1.892106e-04 -0.0002277337 -0.0002603590 -0.0002854800
#>  [4,]  1.539004e-04  1.370133e-04  0.0001603354  0.0001590438  0.0002012006
#>  [5,] -2.877012e-04 -7.233998e-04 -0.0007211156 -0.0007698223 -0.0008312807
#>  [6,]  4.349019e-04  6.865044e-04  0.0007341896  0.0007680830  0.0008950424
#>  [7,]  2.531251e-04  3.221156e-04  0.0003665578  0.0003810441  0.0004583018
#>  [8,]  3.960229e-04  6.031826e-04  0.0006485533  0.0006947873  0.0008552801
#>  [9,] -2.449087e-04 -5.132204e-04 -0.0004840819 -0.0005205386 -0.0005635113
#> [10,] -1.577863e-04 -4.850442e-04 -0.0004771111 -0.0005240404 -0.0005657475
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0006156224 -0.0004024795 -3.078690e-04 -1.732325e-04 -4.042717e-04
#>  [2,]  0.0001845033  0.0001218212  7.648263e-05  4.777007e-05  1.597144e-06
#>  [3,] -0.0001750247 -0.0001265475 -8.817868e-05 -5.742808e-05 -1.107303e-04
#>  [4,]  0.0001982323  0.0001354599  8.512554e-05  5.179149e-05  2.938639e-05
#>  [5,] -0.0005013903 -0.0003287165 -2.578817e-04 -1.546245e-04 -3.412586e-04
#>  [6,]  0.0006808088  0.0004569915  3.190854e-04  1.937395e-04  2.590448e-04
#>  [7,]  0.0003845111  0.0002599672  1.736346e-04  1.058579e-04  1.043047e-04
#>  [8,]  0.0006603171  0.0004266386  3.087110e-04  1.760209e-04  2.533854e-04
#>  [9,] -0.0003543056 -0.0002325765 -1.829550e-04 -1.001860e-04 -2.841251e-04
#> [10,] -0.0003282006 -0.0002178684 -1.717561e-04 -1.021178e-04 -2.394674e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -1.311960e-05 -6.427309e-05 -4.963824e-05 -2.719001e-05 -2.376745e-05
#>  [2,]  2.320188e-05  1.270103e-05  1.084966e-05  6.590395e-06  7.230200e-06
#>  [3,]  2.202098e-06 -1.734833e-05 -9.403827e-06 -4.088884e-06 -3.840143e-06
#>  [4,]  2.233347e-05  1.362160e-05  1.212941e-05  7.888583e-06  8.151352e-06
#>  [5,] -2.209920e-06 -5.613648e-05 -3.856391e-05 -1.916581e-05 -1.646908e-05
#>  [6,]  4.122188e-05  5.959074e-05  4.491379e-05  2.511066e-05  2.427685e-05
#>  [7,]  3.187504e-05  3.029066e-05  2.408678e-05  1.441308e-05  1.441101e-05
#>  [8,]  4.434649e-05  5.873170e-05  4.866376e-05  2.743861e-05  2.728205e-05
#>  [9,] -4.266689e-06 -3.941220e-05 -3.175336e-05 -1.877297e-05 -1.684388e-05
#> [10,]  4.226276e-06 -3.715387e-05 -2.437037e-05 -1.121026e-05 -9.378630e-06
#>                 V55           V56           V57           V58           V59
#>  [1,] -1.008554e-05 -8.181341e-06 -1.320023e-05 -3.155522e-05 -2.877276e-05
#>  [2,]  6.917170e-06  7.954059e-06  1.714388e-06  4.291660e-06  4.346554e-06
#>  [3,]  1.243848e-06 -6.939472e-08  1.797007e-06 -8.368441e-06 -6.607561e-06
#>  [4,]  7.237405e-06  7.722130e-06  3.054203e-06  5.046676e-06  5.273010e-06
#>  [5,] -3.953226e-06 -4.252126e-06 -6.276048e-06 -2.725434e-05 -2.302347e-05
#>  [6,]  1.530883e-05  1.615132e-05  7.800592e-06  2.571561e-05  2.327391e-05
#>  [7,]  1.054673e-05  1.154027e-05  4.685561e-06  1.229022e-05  1.169452e-05
#>  [8,]  1.776304e-05  1.757574e-05  1.346090e-05  2.661359e-05  2.558710e-05
#>  [9,] -7.417935e-06 -5.211531e-06 -1.133095e-05 -2.183822e-05 -2.068368e-05
#> [10,] -5.314466e-07 -7.602776e-07 -2.744442e-06 -1.752820e-05 -1.455505e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -3.216801e-04 -2.154738e-05 -2.972787e-04 -4.098962e-04 -0.0007514836
#>  [2,]  2.974655e-05  3.100718e-06  5.707229e-05  8.950880e-05  0.0001389157
#>  [3,] -7.587305e-05 -3.734529e-06 -6.984363e-05 -1.266426e-04 -0.0002232760
#>  [4,]  4.935657e-05  4.409139e-06  7.696652e-05  9.465654e-05  0.0001462608
#>  [5,] -2.646718e-04 -1.557714e-05 -2.308468e-04 -3.474424e-04 -0.0006383186
#>  [6,]  2.496135e-04  1.685330e-05  2.781656e-04  3.927677e-04  0.0006714541
#>  [7,]  1.168834e-04  8.758325e-06  1.488729e-04  2.066909e-04  0.0003426353
#>  [8,]  2.406643e-04  1.912358e-05  2.637823e-04  3.835580e-04  0.0006903393
#>  [9,] -2.159793e-04 -1.584178e-05 -1.971680e-04 -2.563859e-04 -0.0004685213
#> [10,] -1.793914e-04 -9.706992e-06 -1.493049e-04 -2.329627e-04 -0.0004383713
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.01258137  0.01562832  0.01342332  0.01442456  0.01495866  0.01498208
#> [2,] -0.01258137 -0.01562832 -0.01342332 -0.01442456 -0.01495866 -0.01498208
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01190535  0.01271337  0.01217988  0.01425057
#> [2,] -0.01190535 -0.01271337 -0.01217988 -0.01425057
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.7780983 0.5893117 0.1822929 0.2794449 0.6268264 0.4268321 0.4721648
#>  [8] 0.7867534 0.6519588 0.4550759
#> 
#> $B[[2]]
#> [1]  0.07892361 -0.05337963
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0002417598 0.0006766237 0.0006264709 0.0007755916 0.0007707074
#>  [6] 0.0007899804 0.0008483922 0.0009437742 0.0001744869 0.0007992042
#> 
#> $vB[[2]]
#> [1]  0.006665713 -0.006665713
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#>  [2,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#>  [3,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#>  [4,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#>  [5,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#>  [6,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [7,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#>  [8,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#>  [9,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [10,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [11,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [12,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [13,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [14,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [15,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [16,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [17,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [18,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [19,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [20,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [21,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [22,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [23,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [24,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [25,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [26,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [27,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [28,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [29,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [30,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [31,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [32,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [33,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [34,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [35,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [36,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [37,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [38,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [39,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [40,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#>  [2,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#>  [3,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#>  [4,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#>  [5,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#>  [6,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [7,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#>  [8,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#>  [9,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [10,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [11,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [12,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [13,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [14,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [15,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [16,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [17,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [18,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [19,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [20,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [21,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [22,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [23,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [24,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [25,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [26,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [27,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [28,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [29,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [30,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [31,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [32,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [33,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [34,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [35,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [36,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [37,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [38,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [39,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [40,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#>  [2,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#>  [3,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#>  [4,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#>  [5,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#>  [6,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [7,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#>  [8,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#>  [9,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [10,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [11,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [12,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [13,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [14,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [15,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [16,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [17,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [18,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [19,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [20,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [21,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [22,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [23,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [24,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [25,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [26,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [27,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [28,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [29,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [30,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [31,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [32,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [33,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [34,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [35,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [36,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [37,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [38,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [39,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [40,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#>  [2,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#>  [3,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#>  [4,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#>  [5,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#>  [6,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [7,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#>  [8,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#>  [9,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [10,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [11,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [12,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [13,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [14,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [15,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [16,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [17,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [18,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [19,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [20,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [21,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [22,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [23,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [24,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [25,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [26,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [27,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [28,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [29,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [30,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [31,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [32,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [33,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [34,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [35,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [36,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [37,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [38,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [39,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [40,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#>  [2,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#>  [3,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#>  [4,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#>  [5,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#>  [6,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [7,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#>  [8,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#>  [9,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [10,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [11,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [12,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [13,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [14,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [15,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [16,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [17,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [18,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [19,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [20,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [21,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [22,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [23,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [24,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [25,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [26,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [27,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [28,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [29,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [30,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [31,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [32,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [33,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [34,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [35,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [36,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [37,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [38,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [39,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [40,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#>  [2,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#>  [3,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#>  [4,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#>  [5,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#>  [6,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [7,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#>  [8,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#>  [9,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [10,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [11,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [12,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [13,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [14,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [15,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [16,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [17,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [18,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [19,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [20,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [21,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [22,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [23,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [24,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [25,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [26,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [27,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [28,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [29,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [30,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [31,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [32,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [33,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [34,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [35,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [36,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [37,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [38,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [39,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [40,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8348620 0.8244293 0.3398114 0.6121335 0.7317912 0.6824236 0.6520718
#>  [2,] 0.8180416 0.8581257 0.3662206 0.5911650 0.7308970 0.6501145 0.6413588
#>  [3,] 0.8387420 0.8622831 0.3605928 0.6061719 0.7539804 0.6771818 0.6469437
#>  [4,] 0.8062950 0.8399438 0.3197522 0.5736432 0.6970357 0.6161864 0.6248135
#>  [5,] 0.8476856 0.8415202 0.2897754 0.5432506 0.6584506 0.6250174 0.5701529
#>  [6,] 0.7326279 0.6782602 0.2335782 0.3887854 0.4666032 0.4429357 0.4334661
#>  [7,] 0.8758142 0.8977700 0.3335311 0.6093746 0.7605768 0.6880647 0.6444364
#>  [8,] 0.7878464 0.7813293 0.2212149 0.4416705 0.5963657 0.4961349 0.5147164
#>  [9,] 0.8008918 0.7641585 0.2698453 0.5180802 0.6317289 0.6126717 0.5749743
#> [10,] 0.8522422 0.8748755 0.3282715 0.6345176 0.7661642 0.7098733 0.6686744
#> [11,] 0.8175231 0.8479073 0.3044972 0.5528729 0.6478572 0.5794313 0.5973889
#> [12,] 0.8215173 0.8118671 0.2691058 0.5468860 0.6387672 0.6040167 0.6028696
#> [13,] 0.8286186 0.8393703 0.3309307 0.5812504 0.7232699 0.6465443 0.6332523
#> [14,] 0.8176572 0.7805340 0.3687280 0.5685365 0.6468395 0.6408577 0.6065523
#> [15,] 0.7417837 0.7394458 0.2345216 0.4309153 0.5427601 0.4601280 0.4817165
#> [16,] 0.7572440 0.7685093 0.1842044 0.4172858 0.5326812 0.4463264 0.4501241
#> [17,] 0.8711234 0.8983194 0.3645877 0.6323333 0.7813347 0.7171528 0.6658122
#> [18,] 0.8600789 0.8690350 0.3286779 0.6129626 0.7582202 0.6972060 0.6542863
#> [19,] 0.7190592 0.6128792 0.2971198 0.4500402 0.4802238 0.4867684 0.5418816
#> [20,] 0.8743411 0.8738941 0.3648702 0.6562362 0.7799713 0.7445798 0.6877141
#> [21,] 0.8324631 0.8068901 0.3185608 0.5817819 0.6934659 0.6600152 0.6187416
#> [22,] 0.8113784 0.8104521 0.2456588 0.4602840 0.6151967 0.5306218 0.5594991
#> [23,] 0.8418470 0.8770752 0.3719267 0.6243064 0.7720719 0.7027996 0.6626764
#> [24,] 0.6716762 0.6140632 0.2408323 0.3922188 0.4419446 0.4394877 0.4590592
#> [25,] 0.8459525 0.8421084 0.3262721 0.5508603 0.6850179 0.6334616 0.5828490
#> [26,] 0.8286993 0.8299317 0.2839464 0.5916391 0.6741304 0.6254369 0.6302010
#> [27,] 0.7903445 0.7121126 0.2603110 0.4874416 0.5634388 0.5390431 0.5522327
#> [28,] 0.7698217 0.7892848 0.1983820 0.4072616 0.4862856 0.4426592 0.4842990
#> [29,] 0.8254953 0.7967879 0.4301433 0.6242300 0.7258392 0.6974243 0.6346967
#> [30,] 0.8091034 0.8414489 0.3189931 0.5626032 0.7035343 0.6105754 0.6180705
#> [31,] 0.7801356 0.7885699 0.2499192 0.4167108 0.5069915 0.4562332 0.4817812
#> [32,] 0.8360259 0.8281455 0.3205684 0.5396038 0.6860918 0.6307314 0.6039046
#> [33,] 0.8227050 0.7766688 0.2802166 0.4964107 0.5927625 0.5998308 0.5451293
#> [34,] 0.8199161 0.7915256 0.3329750 0.5413746 0.6390249 0.6044136 0.6036451
#> [35,] 0.8319428 0.8635069 0.3253936 0.5593129 0.6974680 0.6098457 0.6147539
#> [36,] 0.8372920 0.8672039 0.3532538 0.6297192 0.7598087 0.7006436 0.6655959
#> [37,] 0.8405923 0.8521440 0.2467910 0.5097128 0.6217657 0.5650604 0.5367333
#> [38,] 0.7347878 0.7063680 0.2930432 0.4800796 0.5506517 0.5082254 0.5397129
#> [39,] 0.7472184 0.6972476 0.2668836 0.3992952 0.4680231 0.4292168 0.4523281
#> [40,] 0.8200979 0.7612383 0.3144820 0.5345401 0.6250853 0.6108581 0.5651211
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8966835 0.8808786 0.6979300
#>  [2,] 0.8740960 0.8686352 0.6497275
#>  [3,] 0.8991707 0.8875274 0.6846797
#>  [4,] 0.8664203 0.8610868 0.6291124
#>  [5,] 0.8834231 0.8795459 0.6301996
#>  [6,] 0.7465001 0.7881069 0.4778327
#>  [7,] 0.9255992 0.9186813 0.6841626
#>  [8,] 0.8317096 0.8530359 0.5258070
#>  [9,] 0.8648462 0.8479815 0.6602836
#> [10,] 0.9168753 0.9001312 0.7077347
#> [11,] 0.8657522 0.8739026 0.6010125
#> [12,] 0.8735379 0.8648222 0.6603281
#> [13,] 0.8752876 0.8682489 0.6669912
#> [14,] 0.8637611 0.8459736 0.6530584
#> [15,] 0.7657023 0.7999984 0.5040112
#> [16,] 0.7934134 0.8224546 0.4672667
#> [17,] 0.9254040 0.9130283 0.7066821
#> [18,] 0.9165176 0.9076827 0.7113283
#> [19,] 0.7242772 0.7235543 0.5772450
#> [20,] 0.9264398 0.9098281 0.7662280
#> [21,] 0.8907850 0.8774098 0.6821037
#> [22,] 0.8583226 0.8700638 0.5691725
#> [23,] 0.9037195 0.8865701 0.6885847
#> [24,] 0.6856445 0.7165466 0.5034633
#> [25,] 0.9007053 0.8981205 0.6170337
#> [26,] 0.8819984 0.8724224 0.6720755
#> [27,] 0.8241177 0.8181279 0.6211570
#> [28,] 0.8093380 0.8479405 0.4857065
#> [29,] 0.8751540 0.8492877 0.6959243
#> [30,] 0.8662170 0.8597518 0.6322429
#> [31,] 0.8078436 0.8390481 0.5088557
#> [32,] 0.8832611 0.8771229 0.6497898
#> [33,] 0.8703805 0.8702090 0.6204340
#> [34,] 0.8593412 0.8590654 0.6323558
#> [35,] 0.8747777 0.8761674 0.6190110
#> [36,] 0.9058784 0.8832500 0.6841814
#> [37,] 0.8897357 0.8969252 0.5918261
#> [38,] 0.7528088 0.7660104 0.5656299
#> [39,] 0.7479806 0.7918607 0.4666360
#> [40,] 0.8600655 0.8478914 0.6337474
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6276087 0.3723913
#>  [2,] 0.6267122 0.3732878
#>  [3,] 0.6288825 0.3711175
#>  [4,] 0.6243997 0.3756003
#>  [5,] 0.6249541 0.3750459
#>  [6,] 0.6107715 0.3892285
#>  [7,] 0.6310465 0.3689535
#>  [8,] 0.6186610 0.3813390
#>  [9,] 0.6213819 0.3786181
#> [10,] 0.6297350 0.3702650
#> [11,] 0.6236314 0.3763686
#> [12,] 0.6233501 0.3766499
#> [13,] 0.6267882 0.3732118
#> [14,] 0.6233031 0.3766969
#> [15,] 0.6143017 0.3856983
#> [16,] 0.6139424 0.3860576
#> [17,] 0.6317459 0.3682541
#> [18,] 0.6303038 0.3696962
#> [19,] 0.6105453 0.3894547
#> [20,] 0.6325010 0.3674990
#> [21,] 0.6259286 0.3740714
#> [22,] 0.6212708 0.3787292
#> [23,] 0.6294151 0.3705849
#> [24,] 0.6061488 0.3938512
#> [25,] 0.6260550 0.3739450
#> [26,] 0.6251520 0.3748480
#> [27,] 0.6182280 0.3817720
#> [28,] 0.6143679 0.3856321
#> [29,] 0.6269985 0.3730015
#> [30,] 0.6248007 0.3751993
#> [31,] 0.6161066 0.3838934
#> [32,] 0.6257601 0.3742399
#> [33,] 0.6213567 0.3786433
#> [34,] 0.6233528 0.3766472
#> [35,] 0.6259640 0.3740360
#> [36,] 0.6283543 0.3716457
#> [37,] 0.6235637 0.3764363
#> [38,] 0.6144435 0.3855565
#> [39,] 0.6120065 0.3879935
#> [40,] 0.6221141 0.3778859
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]       [,3]        [,4]        [,5]        [,6]
#>  [1,] 1.6204848 1.5466496 -0.6641347  0.45628910  1.00372937  0.76493232
#>  [2,] 1.5031353 1.7998092 -0.5484653  0.36878361  0.99917840  0.61954243
#>  [3,] 1.6488976 1.8343833 -0.5727921  0.43124905  1.11995492  0.74085047
#>  [4,] 1.4261135 1.6578098 -0.7549110  0.29673115  0.83322161  0.47339271
#>  [5,] 1.7165628 1.6695829 -0.8964749  0.17343595  0.65639709  0.51089994
#>  [6,] 1.0079966 0.7457878 -1.1882159 -0.45242068 -0.13378654 -0.22925616
#>  [7,] 1.9533749 2.1726885 -0.6922576  0.44468408  1.15584432  0.79108721
#>  [8,] 1.3119926 1.2734298 -1.2586004 -0.23438503  0.39034461 -0.01546064
#>  [9,] 1.3918772 1.1756153 -0.9954075  0.07235239  0.53964112  0.45855660
#> [10,] 1.7522959 1.9447727 -0.7160134  0.55164701  1.18677749  0.89476854
#> [11,] 1.4996554 1.7182813 -0.8259732  0.21228521  0.60963360  0.32043929
#> [12,] 1.5266615 1.4621880 -0.9991640  0.18809664  0.57001760  0.42222983
#> [13,] 1.5758683 1.6535505 -0.7039784  0.32790863  0.96073972  0.60388339
#> [14,] 1.5005547 1.2687807 -0.5376777  0.27588270  0.60517577  0.57908863
#> [15,] 1.0552600 1.0430900 -1.1829531 -0.27811781  0.17145934 -0.15982726
#> [16,] 1.1376285 1.1999127 -1.4881181 -0.33392560  0.13091154 -0.21552480
#> [17,] 1.9109279 2.1786892 -0.5555070  0.54223975  1.27346148  0.93038224
#> [18,] 1.8159452 1.8924532 -0.7141707  0.45978268  1.14294629  0.83402826
#> [19,] 0.9397997 0.4594309 -0.8610509 -0.20050835 -0.07914607 -0.05293883
#> [20,] 1.9398997 1.9358372 -0.5542877  0.64656589  1.26549895  1.06990998
#> [21,] 1.6031853 1.4299281 -0.7603935  0.33009276  0.81637310  0.66336189
#> [22,] 1.4589914 1.4529507 -1.1219013 -0.15919918  0.46920980  0.12264079
#> [23,] 1.6720353 1.9650203 -0.5239599  0.50786746  1.22004723  0.86066518
#> [24,] 0.7157761 0.4644244 -1.1481216 -0.43799490 -0.23327374 -0.24324147
#> [25,] 1.7032023 1.6739998 -0.7250946  0.20414727  0.77692929  0.54709591
#> [26,] 1.5764367 1.5851432 -0.9249694  0.37074557  0.72692623  0.51269031
#> [27,] 1.3270034 0.9056666 -1.0443530 -0.05024409  0.25512996  0.15649083
#> [28,] 1.2073047 1.3206196 -1.3964375 -0.37529729 -0.05487124 -0.23037669
#> [29,] 1.5540315 1.3663384 -0.2812663  0.50754157  0.97361392  0.83506225
#> [30,] 1.4441949 1.6690485 -0.7584027  0.25173379  0.86418489  0.44973160
#> [31,] 1.2664566 1.3163268 -1.0990433 -0.33629054  0.02796774 -0.17551627
#> [32,] 1.6289513 1.5725403 -0.7511609  0.15874774  0.78191074  0.53535565
#> [33,] 1.5347824 1.2463579 -0.9433873 -0.01435742  0.37539746  0.40476029
#> [34,] 1.5157794 1.3341462 -0.6947601  0.16587772  0.57113429  0.42388934
#> [35,] 1.5994591 1.8447280 -0.7290941  0.23837417  0.83526936  0.44666363
#> [36,] 1.6382156 1.8764596 -0.6047672  0.53101232  1.15163098  0.85036439
#> [37,] 1.6626414 1.7515165 -1.1158010  0.03885617  0.49704932  0.26172534
#> [38,] 1.0190515 0.8778093 -0.8806497 -0.07972360  0.20330430  0.03290469
#> [39,] 1.0838318 0.8342252 -1.0104918 -0.40840261 -0.12808261 -0.28504740
#> [40,] 1.5170108 1.1594806 -0.7792481  0.13838093  0.51118942  0.45092046
#>              [,7]      [,8]      [,9]       [,10]
#>  [1,]  0.62815844 2.1609055 2.0007767  0.83746015
#>  [2,]  0.58126657 1.9376704 1.8889448  0.61784180
#>  [3,]  0.60563168 2.1880439 2.0657300  0.77536197
#>  [4,]  0.51002976 1.8696717 1.8243459  0.52841110
#>  [5,]  0.28247486 2.0252528 1.9881372  0.53307335
#>  [6,] -0.26772333 1.0800324 1.3135517 -0.08872720
#>  [7,]  0.59467193 2.5209751 2.4245637  0.77296817
#>  [8,]  0.05888244 1.5977922 1.7586136  0.10332000
#>  [9,]  0.30217557 1.8561379 1.7188564  0.66455842
#> [10,]  0.70219563 2.4006298 2.1986837  0.88440725
#> [11,]  0.39459719 1.8639115 1.9359139  0.40968570
#> [12,]  0.41743607 1.9326085 1.8559330  0.66475646
#> [13,]  0.54619474 1.9485422 1.8855637  0.69460784
#> [14,]  0.43284275 1.8468859 1.7033645  0.63250998
#> [15,] -0.07316660 1.1842012 1.3862846  0.01604519
#> [16,] -0.20016921 1.3456244 1.5330669 -0.13112057
#> [17,]  0.68930448 2.5181429 2.3511846  0.87932391
#> [18,]  0.63793413 2.3959457 2.2856629  0.90184420
#> [19,]  0.16791989 0.9657779 0.9621611  0.31147413
#> [20,]  0.78945405 2.5332454 2.3115382  1.18713351
#> [21,]  0.48421020 2.0987848 1.9681270  0.76345639
#> [22,]  0.23912952 1.8014273 1.9015229  0.27847559
#> [23,]  0.67524423 2.2392533 2.0561749  0.79351091
#> [24,] -0.16413064 0.7798350 0.9273958  0.01385331
#> [25,]  0.33447984 2.2050863 2.1765136  0.47697666
#> [26,]  0.53307939 2.0114925 1.9225487  0.71758708
#> [27,]  0.20969577 1.5444982 1.5037153  0.49446205
#> [28,] -0.06282464 1.4457144 1.7185389 -0.05718965
#> [29,]  0.55241932 1.9473191 1.7290256  0.82796405
#> [30,]  0.48136660 1.8679162 1.8132299  0.54185098
#> [31,] -0.07290764 1.4360587 1.6511625  0.03542661
#> [32,]  0.42176134 2.0236814 1.9654629  0.61811559
#> [33,]  0.18100997 1.9043276 1.9028083  0.49139094
#> [34,]  0.42067651 1.8098291 1.8075492  0.54233673
#> [35,]  0.46733955 1.9438796 1.9566267  0.48535262
#> [36,]  0.68833239 2.2643180 2.0235734  0.77305509
#> [37,]  0.14719842 2.0880447 2.1635174  0.37151939
#> [38,]  0.15918701 1.1136490 1.1859193  0.26404306
#> [39,] -0.19126867 1.0878708 1.3361780 -0.13365472
#> [40,]  0.26197239 1.8158341 1.7181577  0.54832719
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2434317 -0.2785401
#>  [2,] 0.2415558 -0.2765820
#>  [3,] 0.2462369 -0.2811889
#>  [4,] 0.2359310 -0.2723342
#>  [5,] 0.2395707 -0.2710592
#>  [6,] 0.2231303 -0.2274261
#>  [7,] 0.2489139 -0.2877951
#>  [8,] 0.2335065 -0.2503624
#>  [9,] 0.2360281 -0.2593896
#> [10,] 0.2435280 -0.2875521
#> [11,] 0.2347040 -0.2702864
#> [12,] 0.2351227 -0.2686694
#> [13,] 0.2425725 -0.2758900
#> [14,] 0.2388662 -0.2647259
#> [15,] 0.2248289 -0.2406020
#> [16,] 0.2218301 -0.2420845
#> [17,] 0.2496678 -0.2900465
#> [18,] 0.2480021 -0.2855184
#> [19,] 0.2206228 -0.2289820
#> [20,] 0.2507635 -0.2921978
#> [21,] 0.2420644 -0.2727251
#> [22,] 0.2375668 -0.2573790
#> [23,] 0.2451929 -0.2845154
#> [24,] 0.2127489 -0.2184034
#> [25,] 0.2441399 -0.2711895
#> [26,] 0.2352720 -0.2762025
#> [27,] 0.2325003 -0.2495333
#> [28,] 0.2224275 -0.2432825
#> [29,] 0.2457918 -0.2735698
#> [30,] 0.2381517 -0.2718236
#> [31,] 0.2285797 -0.2444755
#> [32,] 0.2443226 -0.2697474
#> [33,] 0.2379675 -0.2573433
#> [34,] 0.2396685 -0.2641352
#> [35,] 0.2405864 -0.2743544
#> [36,] 0.2413519 -0.2838114
#> [37,] 0.2371516 -0.2675506
#> [38,] 0.2234689 -0.2425603
#> [39,] 0.2252511 -0.2305033
#> [40,] 0.2384859 -0.2600455
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3723913 -0.3723913
#>  [2,] -0.6267122  0.6267122
#>  [3,] -0.6288825  0.6288825
#>  [4,]  0.3756003 -0.3756003
#>  [5,]  0.3750459 -0.3750459
#>  [6,]  0.3892285 -0.3892285
#>  [7,] -0.6310465  0.6310465
#>  [8,]  0.3813390 -0.3813390
#>  [9,]  0.3786181 -0.3786181
#> [10,]  0.3702650 -0.3702650
#> [11,] -0.6236314  0.6236314
#> [12,] -0.6233501  0.6233501
#> [13,] -0.6267882  0.6267882
#> [14,] -0.6233031  0.6233031
#> [15,]  0.3856983 -0.3856983
#> [16,]  0.3860576 -0.3860576
#> [17,] -0.6317459  0.6317459
#> [18,]  0.3696962 -0.3696962
#> [19,]  0.3894547 -0.3894547
#> [20,] -0.6325010  0.6325010
#> [21,]  0.3740714 -0.3740714
#> [22,]  0.3787292 -0.3787292
#> [23,]  0.3705849 -0.3705849
#> [24,]  0.3938512 -0.3938512
#> [25,]  0.3739450 -0.3739450
#> [26,] -0.6251520  0.6251520
#> [27,]  0.3817720 -0.3817720
#> [28,]  0.3856321 -0.3856321
#> [29,] -0.6269985  0.6269985
#> [30,]  0.3751993 -0.3751993
#> [31,] -0.6161066  0.6161066
#> [32,]  0.3742399 -0.3742399
#> [33,] -0.6213567  0.6213567
#> [34,] -0.6233528  0.6233528
#> [35,] -0.6259640  0.6259640
#> [36,]  0.3716457 -0.3716457
#> [37,] -0.6235637  0.6235637
#> [38,]  0.3855565 -0.3855565
#> [39,] -0.6120065  0.6120065
#> [40,]  0.3778859 -0.3778859
#> 
#> $L
#> [1] 0.6924584 0.7050601 0.6927072 0.7116558 0.7079276 0.6906810
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319