Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12          V13          V14
#>  [1,] -0.3027105 -0.05212921 -0.143122783 -0.18829935 -0.045106196 -0.037339211
#>  [2,] -0.2464641 -0.19841683 -0.179616568 -0.13661296 -0.014782460 -0.139937557
#>  [3,] -0.4453908 -0.09290424  0.005103817 -0.01431960 -0.037653007 -0.033037838
#>  [4,] -0.3653395 -0.18066939 -0.028322966 -0.03990443 -0.186674178 -0.064757311
#>  [5,] -0.3788238 -0.03871593 -0.213891242 -0.14493824 -0.002953636  0.007335586
#>  [6,] -0.2550896 -0.09485307 -0.060485739  0.04047063 -0.112074589 -0.055747738
#>  [7,] -0.3393702 -0.18504031 -0.209356290 -0.11120391 -0.029520336 -0.140346709
#>  [8,] -0.2101824 -0.12067778 -0.151374549 -0.11110303 -0.118607721 -0.144018179
#>  [9,] -0.3306873 -0.12103221 -0.032615341 -0.02945086 -0.124793773  0.007826561
#> [10,] -0.2321777 -0.09005119 -0.084751888 -0.06516032 -0.060248689 -0.126417204
#>                V15          V16         V17         V18         V19         V2
#>  [1,] -0.043414706 -0.093744736 -0.02320957  0.07270886  0.05610392 -0.3006084
#>  [2,]  0.009703606 -0.095819341 -0.05890180 -0.04985876 -0.04037347 -0.2063413
#>  [3,] -0.100221676  0.002424555  0.12835876  0.03548060  0.12788302 -0.2863952
#>  [4,] -0.121411874 -0.188045067 -0.02267078 -0.13179311 -0.06793303 -0.3499110
#>  [5,] -0.057385874  0.054019401  0.14341819  0.15813087  0.10549395 -0.2339276
#>  [6,]  0.005272484  0.085412523 -0.00206140  0.12613651  0.08209652 -0.3209251
#>  [7,] -0.023565512 -0.008912463 -0.03735939 -0.06703393  0.03306962 -0.3362083
#>  [8,] -0.107298307 -0.065781323 -0.07740094 -0.01233228  0.18851299 -0.2026419
#>  [9,]  0.038410074  0.111592228  0.16842614  0.18089262  0.11863104 -0.3573823
#> [10,] -0.091420992  0.045585790  0.01899164  0.05811940  0.04434324 -0.3285188
#>                V20         V21        V22          V23         V24          V25
#>  [1,]  0.285280787  0.30105201 0.12761900  0.218605912  0.18962921  0.268331043
#>  [2,]  0.015894716 -0.03998772 0.09126452  0.086324036  0.14669508  0.046875494
#>  [3,]  0.248221405  0.16414051 0.19155044  0.136845676  0.18908143  0.240695486
#>  [4,]  0.020626279 -0.02634583 0.18116748  0.139808279  0.01055866  0.168840620
#>  [5,] -0.004017333  0.01466105 0.02357206  0.130496680  0.17897617  0.016209578
#>  [6,]  0.080483102  0.25869912 0.13411690  0.203015726  0.16714727  0.160951366
#>  [7,] -0.027821805  0.05889556 0.02296215  0.005339737  0.09326127  0.050330561
#>  [8,]  0.111590716  0.11876709 0.16753540  0.135109800  0.07865169  0.216152899
#>  [9,]  0.173101995  0.10539374 0.16502178  0.108074894  0.15027824  0.136079712
#> [10,]  0.009512609  0.09837969 0.11679915 -0.006689857 -0.02676938 -0.006518884
#>               V26        V27         V28           V29         V3           V30
#>  [1,]  0.23311558 0.17536401  0.17649175  0.1244415086 -0.2258529  0.1470329590
#>  [2,]  0.18815315 0.06773652  0.13779470  0.0115110935 -0.2179614 -0.0127787528
#>  [3,]  0.19390669 0.14871535  0.14070639  0.1115638351 -0.2642898  0.1026970825
#>  [4,]  0.17421632 0.18145458  0.13308562  0.0210064120 -0.3919072 -0.0303107290
#>  [5,]  0.07649656 0.05004698  0.09458575  0.0380005402 -0.3775203  0.0646555193
#>  [6,]  0.23791227 0.18881653  0.28953845  0.1119670677 -0.2411923  0.0004659137
#>  [7,] -0.01469534 0.09965202 -0.01760808  0.0001737664 -0.2402099 -0.0648272895
#>  [8,]  0.11081384 0.12882065  0.08419541 -0.0286335706 -0.2589061 -0.0079425974
#>  [9,]  0.27824293 0.17619891  0.06685922  0.1729246537 -0.3307780  0.1219195196
#> [10,]  0.06304898 0.01138447 -0.12062966  0.0393609617 -0.2124674 -0.1137971161
#>               V31          V32          V33         V34          V35
#>  [1,]  0.02480547  0.075065764 -0.010423812  0.07212227  0.041929780
#>  [2,] -0.11509797  0.066557031 -0.077841519  0.05811160 -0.012525298
#>  [3,]  0.04590169  0.081259647  0.053817137 -0.04479309 -0.071680891
#>  [4,] -0.02542688 -0.044218699  0.006683024  0.03898226  0.006183728
#>  [5,]  0.05228507 -0.061399933 -0.055759147  0.02681405  0.049341693
#>  [6,]  0.15095278  0.073680756  0.125732072  0.15415977  0.022022169
#>  [7,] -0.07504008 -0.084285347 -0.055861712  0.05004630 -0.001139711
#>  [8,]  0.07858855 -0.007353354 -0.091794076 -0.08999639  0.023083614
#>  [9,] -0.06790735  0.042890302 -0.027214997 -0.12059609 -0.012919402
#> [10,] -0.01557340  0.017429004 -0.154816509 -0.08225774 -0.063233695
#>               V36         V37         V38         V39         V4           V40
#>  [1,] -0.01845423 -0.03813474  0.04069592  0.01472464 -0.2362753 -0.1240450834
#>  [2,]  0.07382946  0.07282309  0.07804470  0.03604603 -0.2418840 -0.1086222513
#>  [3,] -0.10221374 -0.08197885 -0.05261817  0.01216643 -0.3724125 -0.0092775554
#>  [4,] -0.16384133 -0.08396491 -0.15812684 -0.02707545 -0.2820148 -0.0128307856
#>  [5,]  0.04567081  0.10605830 -0.12750138 -0.04975175 -0.3137529 -0.0717058273
#>  [6,]  0.02069677 -0.04341040 -0.08851724  0.07152563 -0.2463059  0.0209941709
#>  [7,]  0.04358226  0.01360219 -0.16619942 -0.19913955 -0.3048707 -0.1495904518
#>  [8,] -0.10096319 -0.07758687 -0.03475377 -0.10876606 -0.2380519 -0.0860583671
#>  [9,]  0.06161454 -0.04346464  0.02041082 -0.10336983 -0.1996276 -0.0886279073
#> [10,]  0.07154812  0.01383273 -0.09942384 -0.02624168 -0.1642566  0.0009206931
#>                V41          V42         V43          V44         V45
#>  [1,] -0.035739721 -0.042253013 -0.09658908 -0.095333113 -0.15991827
#>  [2,] -0.162896133 -0.123006938 -0.15037927 -0.128259855 -0.14272814
#>  [3,] -0.124595682 -0.006965473 -0.16802399 -0.147761294 -0.07903231
#>  [4,] -0.087370289 -0.024066832 -0.11527220 -0.128451084 -0.10238154
#>  [5,] -0.063631533 -0.148899669 -0.02335462 -0.192262050 -0.21948411
#>  [6,] -0.061874325 -0.031384252 -0.07431325  0.004478601 -0.02260242
#>  [7,] -0.127674397 -0.136863399 -0.06375573 -0.115195434 -0.11352612
#>  [8,] -0.029593519 -0.017389870 -0.02190946 -0.198147232 -0.16932579
#>  [9,] -0.033367140 -0.081082928 -0.03974799 -0.024781557 -0.11802978
#> [10,] -0.001955869 -0.087690719 -0.05947725 -0.120381040 -0.18473180
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.05173928 -0.09151411 -0.2016617 -0.3832620 -0.2146385 -0.3476580
#>  [2,] -0.05583784 -0.18236970 -0.2400610 -0.2753851 -0.2307281 -0.2973275
#>  [3,] -0.16670503 -0.22906980 -0.3399237 -0.3404666 -0.2981647 -0.2902648
#>  [4,] -0.10564073 -0.27065270 -0.1858423 -0.2323103 -0.2494619 -0.2805325
#>  [5,] -0.25098414 -0.16504933 -0.2202355 -0.2177810 -0.1926782 -0.2906593
#>  [6,] -0.10775221 -0.19435307 -0.2706640 -0.2973583 -0.1619500 -0.3876786
#>  [7,] -0.20741382 -0.19634839 -0.1219466 -0.2015866 -0.3031479 -0.2518515
#>  [8,] -0.18115012 -0.12222512 -0.2178215 -0.2581557 -0.1953021 -0.2606853
#>  [9,] -0.18515589 -0.20161327 -0.2514011 -0.2562661 -0.2587462 -0.2747868
#> [10,] -0.21506841 -0.20276175 -0.1316941 -0.3192687 -0.1853352 -0.3528352
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2418722 -0.2948966 -0.2889005 -0.3488167 -0.3726561 -0.2577678
#>  [2,] -0.2769816 -0.2783969 -0.3119795 -0.3416015 -0.2464466 -0.3483092
#>  [3,] -0.4282546 -0.4116512 -0.4443759 -0.3213673 -0.4785270 -0.4771053
#>  [4,] -0.3756845 -0.4354739 -0.3660507 -0.3580102 -0.3274866 -0.2697228
#>  [5,] -0.3644878 -0.3414243 -0.2971253 -0.4261519 -0.2558310 -0.3477268
#>  [6,] -0.2938305 -0.2937341 -0.2511691 -0.3862303 -0.2567911 -0.2571456
#>  [7,] -0.3392261 -0.3701076 -0.2443033 -0.3311898 -0.3871369 -0.2967925
#>  [8,] -0.2951697 -0.2755626 -0.2194446 -0.2137086 -0.2199626 -0.3659167
#>  [9,] -0.3493644 -0.3199719 -0.4030777 -0.3585213 -0.2925119 -0.4102558
#> [10,] -0.3632199 -0.3525644 -0.3813575 -0.3773449 -0.2192796 -0.3239157
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.3289845 -0.3657255 -0.3026352 -0.2778785 -0.2859524 -0.1858298
#>  [2,] -0.2149254 -0.3304818 -0.2101292 -0.2087156 -0.3278105 -0.2476734
#>  [3,] -0.4243293 -0.4327511 -0.3210249 -0.1990173 -0.3646181 -0.1640894
#>  [4,] -0.3144909 -0.2817714 -0.2599744 -0.2798289 -0.3700240 -0.2150724
#>  [5,] -0.4131732 -0.3490166 -0.4187934 -0.1476673 -0.2733592 -0.2336870
#>  [6,] -0.3233480 -0.2480188 -0.3835587 -0.2359490 -0.2846644 -0.2604255
#>  [7,] -0.3203591 -0.3554570 -0.4091213 -0.2156148 -0.4233267 -0.1283020
#>  [8,] -0.2119446 -0.2108304 -0.4044803 -0.1542708 -0.3658506 -0.2349428
#>  [9,] -0.4245117 -0.4034206 -0.2566723 -0.2578761 -0.3688831 -0.2182524
#> [10,] -0.3593216 -0.3763882 -0.2778638 -0.1623270 -0.2875536 -0.1274620
#>                V8          V9
#>  [1,] -0.13678755 -0.14034297
#>  [2,] -0.21511198 -0.23613920
#>  [3,] -0.20450460 -0.04613344
#>  [4,] -0.12505404 -0.04938013
#>  [5,] -0.22044533 -0.10213714
#>  [6,] -0.22685656 -0.15032307
#>  [7,] -0.14443835 -0.22800573
#>  [8,] -0.08435121 -0.10523311
#>  [9,] -0.18675644 -0.08251609
#> [10,] -0.10777917 -0.16799802
#> 
#> $W[[2]]
#>            [,1]         [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.10735194 -0.009542113  0.03065820 -0.00203209 -0.07724984  0.06325381
#> [2,] 0.01141646  0.117017427 -0.02959293  0.01147488  0.05760081 -0.15840359
#>             [,7]       [,8]        [,9]       [,10]
#> [1,]  0.01875210 0.04255559 -0.03722507 -0.04341007
#> [2,] -0.00135527 0.03478846 -0.12001682 -0.03083188
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1          V10           V11           V12          V13
#>  [1,]  1.817574e-05 5.205362e-04  0.0004626896  0.0004233130 6.706199e-04
#>  [2,]  1.359522e-04 8.807333e-04  0.0012840170  0.0013035246 1.015682e-03
#>  [3,]  4.270499e-05 6.202502e-04  0.0006234884  0.0005711505 7.519136e-04
#>  [4,]  6.090658e-05 5.947738e-04  0.0006930206  0.0006738817 7.195680e-04
#>  [5,]  1.758099e-04 1.177773e-03  0.0016281972  0.0016035837 1.304772e-03
#>  [6,] -6.309704e-05 1.061982e-04 -0.0001215296 -0.0001186929 3.066321e-04
#>  [7,] -1.593912e-05 6.432695e-06 -0.0001484844 -0.0001670695 6.736365e-05
#>  [8,]  6.259609e-05 6.766864e-04  0.0007473746  0.0007129623 8.081444e-04
#>  [9,]  2.021434e-05 5.058200e-04  0.0004540499  0.0004178704 6.560919e-04
#> [10,]  3.060952e-05 3.210249e-04  0.0003457815  0.0003392658 4.124383e-04
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0006452778 0.0007825699 0.0009026938 0.0009667971 0.0009404512
#>  [2,] 0.0007497857 0.0004847418 0.0004461335 0.0004969081 0.0007821295
#>  [3,] 0.0006957968 0.0007663483 0.0008538841 0.0008933143 0.0009125327
#>  [4,] 0.0006385171 0.0005982032 0.0006542683 0.0007126700 0.0008142906
#>  [5,] 0.0010059886 0.0007009772 0.0006827467 0.0007301035 0.0010301165
#>  [6,] 0.0003400073 0.0006131652 0.0007591331 0.0008516966 0.0007330688
#>  [7,] 0.0002244061 0.0003725322 0.0004906060 0.0005296026 0.0004595949
#>  [8,] 0.0007350096 0.0007262119 0.0008005475 0.0008577890 0.0009380657
#>  [9,] 0.0006309880 0.0007487812 0.0008488714 0.0009052594 0.0009123967
#> [10,] 0.0004098478 0.0004257928 0.0004912865 0.0005425754 0.0005919605
#>                V19           V2           V20           V21           V22
#>  [1,] 0.0008187602 7.244414e-05  7.962432e-04  8.490866e-04  7.499640e-04
#>  [2,] 0.0015180531 1.357895e-04  1.900225e-03  2.024076e-03  1.960149e-03
#>  [3,] 0.0009572581 8.896700e-05  1.040115e-03  1.122234e-03  9.890660e-04
#>  [4,] 0.0010252501 8.757019e-05  1.172950e-03  1.252121e-03  1.179533e-03
#>  [5,] 0.0018821804 1.829844e-04  2.390003e-03  2.579049e-03  2.432751e-03
#>  [6,] 0.0002467197 5.585264e-06 -4.797691e-05 -5.930685e-05 -8.577714e-06
#>  [7,] 0.0002034556 3.809982e-06  1.424412e-04  1.602781e-04  1.761994e-04
#>  [8,] 0.0011119875 9.880007e-05  1.261365e-03  1.348027e-03  1.239264e-03
#>  [9,] 0.0008557921 7.114001e-05  8.663147e-04  9.283872e-04  8.359270e-04
#> [10,] 0.0006479883 4.871808e-05  7.081406e-04  7.526803e-04  7.310221e-04
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0007545531 0.0011058328 0.0012844717 0.0013322244 0.0014069682
#>  [2,] 0.0015003073 0.0011324153 0.0008992375 0.0008909738 0.0008850235
#>  [3,] 0.0008894942 0.0011466820 0.0012734815 0.0013454331 0.0014032204
#>  [4,] 0.0010364080 0.0011207998 0.0011623622 0.0012185354 0.0012492582
#>  [5,] 0.0018618373 0.0015002261 0.0012769711 0.0013411078 0.0013465296
#>  [6,] 0.0003114718 0.0008917332 0.0011849739 0.0011883851 0.0012765513
#>  [7,] 0.0004010899 0.0007271694 0.0009293459 0.0010056529 0.0009889416
#>  [8,] 0.0010947950 0.0012511952 0.0013327076 0.0014071215 0.0014563950
#>  [9,] 0.0008344849 0.0011520610 0.0013039075 0.0013608080 0.0014218874
#> [10,] 0.0007122546 0.0008200828 0.0008840543 0.0009221768 0.0009254399
#>                V28          V29           V3          V30          V31
#>  [1,] 0.0015309800 0.0012185447 9.988436e-05 0.0010935914 0.0011904440
#>  [2,] 0.0010961772 0.0012801758 1.390631e-04 0.0011642452 0.0002996819
#>  [3,] 0.0015348480 0.0012890745 1.152271e-04 0.0011738702 0.0010898114
#>  [4,] 0.0013167156 0.0011776683 1.062935e-04 0.0010610413 0.0008299771
#>  [5,] 0.0015951627 0.0017413475 1.887552e-04 0.0016073024 0.0005783316
#>  [6,] 0.0013349962 0.0009005875 3.468957e-05 0.0007144933 0.0012750283
#>  [7,] 0.0008317479 0.0005688558 2.096688e-05 0.0005017021 0.0008250554
#>  [8,] 0.0015437252 0.0013428299 1.226038e-04 0.0012222065 0.0010284609
#>  [9,] 0.0015211709 0.0012390470 9.945478e-05 0.0010981487 0.0011575463
#> [10,] 0.0009128322 0.0007770811 6.355863e-05 0.0006960067 0.0006463342
#>                V32          V33          V34           V35           V36
#>  [1,] 0.0011812840 0.0010352169 0.0010663027  0.0011417860  0.0010589802
#>  [2,] 0.0002694194 0.0004774987 0.0002546292 -0.0001019557 -0.0004020018
#>  [3,] 0.0010952513 0.0010379118 0.0010506651  0.0010358858  0.0008793554
#>  [4,] 0.0008160482 0.0008222398 0.0007745330  0.0006727298  0.0005018688
#>  [5,] 0.0005517694 0.0007952314 0.0005507056  0.0001203257 -0.0002781143
#>  [6,] 0.0012748160 0.0009583671 0.0010342184  0.0012928196  0.0014307096
#>  [7,] 0.0007872800 0.0006476921 0.0007395407  0.0009027143  0.0009822438
#>  [8,] 0.0010144155 0.0009938716 0.0009663727  0.0008852158  0.0006953764
#>  [9,] 0.0011599118 0.0010641546 0.0010984136  0.0011365202  0.0010421291
#> [10,] 0.0006246072 0.0005937431 0.0005828403  0.0005747141  0.0005114542
#>                 V37          V38          V39            V4          V40
#>  [1,]  8.111633e-04 0.0008966365 0.0007134355  7.287139e-05 0.0005776522
#>  [2,] -1.594410e-04 0.0003342749 0.0006354644  2.707897e-04 0.0004605478
#>  [3,]  6.772621e-04 0.0008577278 0.0007248839  1.175182e-04 0.0005525401
#>  [4,]  4.163839e-04 0.0006236816 0.0006143788  1.434394e-04 0.0004817168
#>  [5,] -5.830079e-05 0.0005569168 0.0008853897  3.516386e-04 0.0006248593
#>  [6,]  1.068025e-03 0.0008835398 0.0005669008 -8.064793e-05 0.0005633454
#>  [7,]  7.420944e-04 0.0005056824 0.0003011821 -2.533833e-05 0.0003213540
#>  [8,]  5.574251e-04 0.0007773405 0.0007210532  1.529373e-04 0.0005560351
#>  [9,]  7.942506e-04 0.0008817598 0.0006997985  7.563732e-05 0.0005676527
#> [10,]  4.245184e-04 0.0004646778 0.0004156533  7.305891e-05 0.0003575074
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0005028992 4.732235e-04  3.506270e-04  2.300164e-04  0.0001313060
#>  [2,] 0.0008786165 1.070418e-03  9.890857e-04  9.177680e-04  0.0012462830
#>  [3,] 0.0005858851 6.076778e-04  4.877967e-04  3.753435e-04  0.0003493323
#>  [4,] 0.0005874051 6.485440e-04  5.462468e-04  4.646714e-04  0.0005181820
#>  [5,] 0.0011304756 1.389862e-03  1.274053e-03  1.180195e-03  0.0015988889
#>  [6,] 0.0002400794 6.168184e-05 -6.267520e-05 -1.846428e-04 -0.0004971227
#>  [7,] 0.0001060412 9.650847e-07 -3.177021e-05 -6.086823e-05 -0.0002416543
#>  [8,] 0.0006416640 6.988354e-04  5.812767e-04  4.855528e-04  0.0005189678
#>  [9,] 0.0005135557 4.860955e-04  3.648024e-04  2.520120e-04  0.0001391617
#> [10,] 0.0003530345 3.485007e-04  2.896142e-04  2.389184e-04  0.0002220842
#>                 V46           V47           V48           V49           V5
#>  [1,]  8.721517e-06  7.792135e-05  7.714073e-05  4.000991e-05 1.735015e-04
#>  [2,]  1.127832e-03  6.896643e-04  5.313391e-04  3.235606e-04 2.554332e-04
#>  [3,]  2.126095e-04  1.863650e-04  1.687615e-04  9.392456e-05 1.985891e-04
#>  [4,]  3.938640e-04  2.719153e-04  2.309846e-04  1.318049e-04 1.895772e-04
#>  [5,]  1.415643e-03  8.594784e-04  6.804362e-04  4.077823e-04 3.350881e-04
#>  [6,] -5.580019e-04 -2.263540e-04 -1.802889e-04 -1.087815e-04 6.200861e-05
#>  [7,] -2.841659e-04 -1.428554e-04 -9.468728e-05 -6.918166e-05 2.572933e-05
#>  [8,]  3.684467e-04  2.655169e-04  2.329768e-04  1.305776e-04 2.146153e-04
#>  [9,]  1.620683e-05  7.883927e-05  8.323055e-05  4.501615e-05 1.709724e-04
#> [10,]  1.507309e-04  1.199016e-04  1.039602e-04  5.611888e-05 1.095509e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  3.379472e-05  1.656868e-05  2.458067e-05  1.628225e-05  1.719885e-05
#>  [2,]  8.335443e-05  9.365487e-05  5.509704e-05  4.286141e-05  4.296040e-05
#>  [3,]  4.508720e-05  3.299079e-05  3.213187e-05  2.257600e-05  2.388677e-05
#>  [4,]  4.701539e-05  4.010915e-05  3.375135e-05  2.485899e-05  2.551248e-05
#>  [5,]  1.096312e-04  1.192949e-04  7.308367e-05  5.566943e-05  5.649861e-05
#>  [6,] -1.575460e-06 -2.980083e-05 -2.697618e-06 -5.014673e-06 -5.132742e-06
#>  [7,] -1.963572e-06 -1.747225e-05 -2.968679e-07 -5.814582e-08  1.143638e-08
#>  [8,]  5.063585e-05  4.114970e-05  3.717845e-05  2.682196e-05  2.776476e-05
#>  [9,]  3.359118e-05  1.870062e-05  2.421949e-05  1.707391e-05  1.821159e-05
#> [10,]  2.542807e-05  1.769826e-05  1.808992e-05  1.365395e-05  1.389578e-05
#>                V55          V56          V57           V58           V59
#>  [1,] 1.951691e-05 1.883146e-05 1.666804e-05  1.028444e-05  1.323126e-05
#>  [2,] 2.110204e-05 2.153732e-05 8.793965e-06  3.483911e-05  3.086277e-05
#>  [3,] 2.087843e-05 2.131867e-05 1.681202e-05  1.551602e-05  1.836326e-05
#>  [4,] 1.825191e-05 1.877791e-05 1.321702e-05  1.693352e-05  1.885113e-05
#>  [5,] 2.967836e-05 2.982754e-05 1.385686e-05  4.318312e-05  4.134804e-05
#>  [6,] 1.132237e-05 8.468562e-06 1.222304e-05 -6.004598e-06 -3.907759e-06
#>  [7,] 6.937391e-06 6.472451e-06 1.099625e-05 -3.228382e-06  1.170976e-06
#>  [8,] 2.134456e-05 2.188477e-05 1.622194e-05  1.795874e-05  2.085252e-05
#>  [9,] 1.858070e-05 1.880792e-05 1.654663e-05  1.128530e-05  1.356411e-05
#> [10,] 1.161289e-05 1.170569e-05 9.935592e-06  8.920339e-06  1.062300e-05
#>                 V6           V60           V7           V8           V9
#>  [1,] 0.0003592023  5.800523e-06 0.0004486331 0.0004560499 0.0005041446
#>  [2,] 0.0001111750  2.845406e-05 0.0001117779 0.0003135488 0.0007041756
#>  [3,] 0.0003340002  1.115240e-05 0.0004165094 0.0004536790 0.0005893464
#>  [4,] 0.0002296186  1.378142e-05 0.0002789065 0.0003471305 0.0005306602
#>  [5,] 0.0001748153  3.723497e-05 0.0001900195 0.0004423831 0.0009612692
#>  [6,] 0.0003696713 -1.080096e-05 0.0004831914 0.0003768630 0.0001608357
#>  [7,] 0.0001563985 -1.799549e-06 0.0001988851 0.0001337863 0.0000513791
#>  [8,] 0.0002909509  1.443754e-05 0.0003555182 0.0004217196 0.0006169667
#>  [9,] 0.0003511030  6.483201e-06 0.0004414705 0.0004397110 0.0004983882
#> [10,] 0.0001610170  7.374851e-06 0.0001962491 0.0002164515 0.0002903044
#> 
#> $vW[[2]]
#>              [,1]        [,2]         [,3]       [,4]        [,5]         [,6]
#> [1,]  0.006488998  0.01721691  0.006686759  0.0109518  0.01626645  0.007496867
#> [2,] -0.006488998 -0.01721691 -0.006686759 -0.0109518 -0.01626645 -0.007496867
#>             [,7]         [,8]         [,9]       [,10]
#> [1,]  0.01687075  0.008856472  0.007976048  0.01242759
#> [2,] -0.01687075 -0.008856472 -0.007976048 -0.01242759
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4022133 0.1825423 0.5761221 0.3549752 0.3720502 0.6014244 0.2129581
#>  [8] 0.4194040 0.5600241 0.1619545
#> 
#> $B[[2]]
#> [1]  0.12128083 -0.01994706
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.002080116 0.001447567 0.002094202 0.001787827 0.002107778 0.001741741
#>  [7] 0.001244027 0.002101803 0.002088546 0.001295930
#> 
#> $vB[[2]]
#> [1]  0.006493498 -0.006493498
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [2,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#>  [3,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#>  [4,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#>  [5,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [6,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#>  [7,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#>  [8,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#>  [9,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [10,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [11,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [12,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [13,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [14,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [15,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [16,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [17,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [18,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [19,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [20,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [21,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [22,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [23,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [24,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [25,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [26,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [27,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [28,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [29,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [30,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [31,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [32,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [33,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [34,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [35,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [36,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [37,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [38,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [39,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [40,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [2,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#>  [3,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#>  [4,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#>  [5,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [6,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#>  [7,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#>  [8,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#>  [9,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [10,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [11,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [12,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [13,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [14,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [15,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [16,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [17,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [18,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [19,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [20,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [21,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [22,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [23,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [24,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [25,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [26,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [27,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [28,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [29,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [30,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [31,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [32,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [33,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [34,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [35,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [36,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [37,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [38,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [39,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [40,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [2,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#>  [3,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#>  [4,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#>  [5,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [6,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#>  [7,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#>  [8,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#>  [9,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [10,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [11,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [12,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [13,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [14,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [15,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [16,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [17,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [18,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [19,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [20,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [21,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [22,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [23,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [24,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [25,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [26,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [27,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [28,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [29,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [30,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [31,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [32,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [33,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [34,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [35,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [36,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [37,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [38,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [39,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [40,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [2,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#>  [3,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#>  [4,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#>  [5,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [6,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#>  [7,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#>  [8,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#>  [9,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [10,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [11,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [12,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [13,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [14,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [15,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [16,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [17,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [18,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [19,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [20,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [21,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [22,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [23,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [24,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [25,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [26,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [27,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [28,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [29,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [30,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [31,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [32,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [33,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [34,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [35,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [36,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [37,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [38,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [39,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [40,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [2,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#>  [3,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#>  [4,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#>  [5,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [6,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#>  [7,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#>  [8,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#>  [9,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [10,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [11,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [12,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [13,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [14,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [15,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [16,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [17,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [18,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [19,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [20,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [21,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [22,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [23,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [24,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [25,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [26,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [27,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [28,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [29,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [30,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [31,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [32,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [33,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [34,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [35,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [36,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [37,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [38,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [39,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [40,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [2,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#>  [3,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#>  [4,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#>  [5,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [6,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#>  [7,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#>  [8,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#>  [9,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [10,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [11,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [12,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [13,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [14,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [15,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [16,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [17,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [18,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [19,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [20,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [21,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [22,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [23,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [24,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [25,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [26,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [27,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [28,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [29,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [30,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [31,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [32,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [33,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [34,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [35,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [36,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [37,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [38,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [39,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [40,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6738416 0.4296829 0.6285685 0.4081337 0.5146130 0.7648388 0.3056020
#>  [2,] 0.8205368 0.4792123 0.8087517 0.4719323 0.6525627 0.8561523 0.3884122
#>  [3,] 0.8469968 0.5459129 0.8243037 0.5939047 0.6289112 0.8674658 0.4415427
#>  [4,] 0.8580587 0.5881572 0.8448369 0.6346669 0.6623913 0.8827149 0.4509873
#>  [5,] 0.8690192 0.5804782 0.8612470 0.6539877 0.6696132 0.8884669 0.4514880
#>  [6,] 0.5634113 0.3515114 0.5041611 0.3078656 0.4054659 0.6751618 0.2141433
#>  [7,] 0.8256444 0.5046800 0.8295426 0.5735924 0.5879740 0.8781332 0.3465109
#>  [8,] 0.6901250 0.3234212 0.6331536 0.3261247 0.4404588 0.7601433 0.2136909
#>  [9,] 0.7930839 0.4078068 0.8227481 0.4395701 0.6342013 0.8479894 0.3432704
#> [10,] 0.7923521 0.5299211 0.7497444 0.5161482 0.5947087 0.8397608 0.3920570
#> [11,] 0.8221941 0.5378265 0.8150135 0.6124374 0.6232815 0.8587898 0.4097260
#> [12,] 0.8152334 0.5295242 0.8050425 0.5983581 0.6074804 0.8585799 0.4046000
#> [13,] 0.8054293 0.4312784 0.8247711 0.4709975 0.6465046 0.8476776 0.3679061
#> [14,] 0.7109185 0.4250637 0.6818539 0.4650206 0.4902688 0.7741201 0.2932808
#> [15,] 0.8247243 0.4400226 0.8414152 0.4928517 0.6572454 0.8720655 0.3576640
#> [16,] 0.8824826 0.5455433 0.8782844 0.6124455 0.6881037 0.9055013 0.4211030
#> [17,] 0.8451535 0.4865946 0.8463579 0.5168365 0.6729625 0.8768818 0.3963040
#> [18,] 0.7434332 0.4561460 0.7670048 0.4762386 0.6186848 0.8171660 0.3484533
#> [19,] 0.8131605 0.5796559 0.7942281 0.5916995 0.6583562 0.8585107 0.4398827
#> [20,] 0.8126364 0.5654189 0.7667768 0.5972657 0.6394823 0.8624939 0.4375310
#> [21,] 0.8875449 0.6407862 0.8644274 0.6779072 0.7075805 0.9051006 0.4896868
#> [22,] 0.7263793 0.4333731 0.6715538 0.4482793 0.4885171 0.8143965 0.2645360
#> [23,] 0.8602881 0.6006044 0.8557814 0.6527797 0.6804649 0.8825087 0.4768287
#> [24,] 0.7881279 0.4690273 0.7858931 0.4284905 0.6606066 0.8587196 0.3364321
#> [25,] 0.7327178 0.3271978 0.7547647 0.3523055 0.5387375 0.8054730 0.2328797
#> [26,] 0.8708989 0.5647288 0.8632908 0.6533756 0.6448433 0.8913228 0.4452040
#> [27,] 0.8192300 0.4980624 0.8029140 0.4948954 0.6475411 0.8543670 0.3929584
#> [28,] 0.8143377 0.5314109 0.8027435 0.5924711 0.6177330 0.8538340 0.4196749
#> [29,] 0.8220413 0.5763842 0.7662622 0.5583623 0.6061693 0.8567043 0.3915556
#> [30,] 0.8826392 0.5433133 0.8802303 0.5995214 0.6889708 0.9026749 0.4167102
#> [31,] 0.8255936 0.4993895 0.8256402 0.5213042 0.6677477 0.8729873 0.3848097
#> [32,] 0.8548832 0.5002543 0.8567503 0.5594610 0.6799021 0.8813406 0.4231070
#> [33,] 0.8848776 0.5631929 0.8826532 0.6151846 0.7185352 0.9046864 0.4729177
#> [34,] 0.8319850 0.5160384 0.8194335 0.6026722 0.6166681 0.8744454 0.3965829
#> [35,] 0.7537618 0.4568937 0.7034622 0.4671158 0.5151909 0.8146815 0.3010480
#> [36,] 0.7800785 0.4259337 0.7590100 0.4201220 0.5664815 0.8329474 0.2895717
#> [37,] 0.8646099 0.5286756 0.8644534 0.5807765 0.6805683 0.8894335 0.4098209
#> [38,] 0.7424722 0.4237334 0.7452629 0.4424680 0.5704351 0.8059010 0.3294578
#> [39,] 0.8181925 0.5913629 0.8026395 0.6462851 0.6698716 0.8683659 0.4835533
#> [40,] 0.8178291 0.5570159 0.8307924 0.5837722 0.6942449 0.8680508 0.4631173
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4793065 0.6555645 0.3541482
#>  [2,] 0.6366396 0.8193509 0.4514424
#>  [3,] 0.6980432 0.8078576 0.4720964
#>  [4,] 0.7034293 0.8332217 0.4544878
#>  [5,] 0.7269398 0.8446506 0.4606859
#>  [6,] 0.3551828 0.5257235 0.2840793
#>  [7,] 0.6242362 0.8073934 0.3978257
#>  [8,] 0.4126784 0.6548833 0.3008210
#>  [9,] 0.5990079 0.8328098 0.4231411
#> [10,] 0.5788342 0.7562106 0.3943073
#> [11,] 0.6585479 0.7928503 0.4160593
#> [12,] 0.6338845 0.7839452 0.4024700
#> [13,] 0.6323441 0.8384895 0.4496231
#> [14,] 0.5064646 0.6496019 0.3362362
#> [15,] 0.6235066 0.8498894 0.4298675
#> [16,] 0.7183011 0.8708782 0.4443220
#> [17,] 0.6763255 0.8553582 0.4647043
#> [18,] 0.5523565 0.7729076 0.4086839
#> [19,] 0.6590030 0.7966093 0.4750261
#> [20,] 0.6261439 0.7769849 0.4092867
#> [21,] 0.7305065 0.8502566 0.4776864
#> [22,] 0.4722248 0.6722857 0.2833719
#> [23,] 0.7182781 0.8476551 0.4839165
#> [24,] 0.5694848 0.8181974 0.4022118
#> [25,] 0.5142175 0.7493759 0.3446380
#> [26,] 0.7182639 0.8411765 0.4401115
#> [27,] 0.6445492 0.8168223 0.4580787
#> [28,] 0.6300968 0.7774972 0.4000211
#> [29,] 0.6133323 0.7782425 0.4119050
#> [30,] 0.7184804 0.8794349 0.4638276
#> [31,] 0.6387608 0.8315836 0.4463936
#> [32,] 0.6860935 0.8577692 0.4658658
#> [33,] 0.7438343 0.8777371 0.5014795
#> [34,] 0.6654183 0.8019915 0.4182367
#> [35,] 0.5536332 0.6981705 0.3386292
#> [36,] 0.5508537 0.7253702 0.3735055
#> [37,] 0.6978867 0.8654871 0.4505756
#> [38,] 0.5353042 0.7527905 0.3716051
#> [39,] 0.6611984 0.8008397 0.4423916
#> [40,] 0.6730730 0.8381719 0.4828651
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6139203 0.3860797
#>  [2,] 0.6277012 0.3722988
#>  [3,] 0.6294633 0.3705367
#>  [4,] 0.6299960 0.3700040
#>  [5,] 0.6313293 0.3686707
#>  [6,] 0.6028061 0.3971939
#>  [7,] 0.6295673 0.3704327
#>  [8,] 0.6163633 0.3836367
#>  [9,] 0.6281418 0.3718582
#> [10,] 0.6235117 0.3764883
#> [11,] 0.6273965 0.3726035
#> [12,] 0.6271091 0.3728909
#> [13,] 0.6284511 0.3715489
#> [14,] 0.6169430 0.3830570
#> [15,] 0.6301968 0.3698032
#> [16,] 0.6332503 0.3667497
#> [17,] 0.6307620 0.3692380
#> [18,] 0.6219627 0.3780373
#> [19,] 0.6256495 0.3743505
#> [20,] 0.6252852 0.3747148
#> [21,] 0.6312964 0.3687036
#> [22,] 0.6190443 0.3809557
#> [23,] 0.6303950 0.3696050
#> [24,] 0.6253418 0.3746582
#> [25,] 0.6228741 0.3771259
#> [26,] 0.6322264 0.3677736
#> [27,] 0.6272929 0.3727071
#> [28,] 0.6264977 0.3735023
#> [29,] 0.6250608 0.3749392
#> [30,] 0.6333321 0.3666679
#> [31,] 0.6285855 0.3714145
#> [32,] 0.6315102 0.3684898
#> [33,] 0.6333647 0.3666353
#> [34,] 0.6292870 0.3707130
#> [35,] 0.6209140 0.3790860
#> [36,] 0.6234794 0.3765206
#> [37,] 0.6317879 0.3682121
#> [38,] 0.6218143 0.3781857
#> [39,] 0.6266153 0.3733847
#> [40,] 0.6278987 0.3721013
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]         [,2]       [,3]        [,4]        [,5]      [,6]
#>  [1,] 0.7256120 -0.283144979 0.52608038 -0.37168594  0.05846867 1.1793938
#>  [2,] 1.5199887 -0.083198803 1.44191911 -0.11238890  0.63032292 1.7836931
#>  [3,] 1.7112379  0.184170466 1.54578205  0.38013095  0.52754868 1.8787352
#>  [4,] 1.7992585  0.356352410 1.69466669  0.55229071  0.67396910 2.0183950
#>  [5,] 1.8923141  0.324736566 1.82568612  0.63661414  0.70643595 2.0751755
#>  [6,] 0.2550182 -0.612402094 0.01664495 -0.81011666 -0.38274120 0.7316251
#>  [7,] 1.5550673  0.018720601 1.58238919  0.29652348  0.35559622 1.9748697
#>  [8,] 0.8007036 -0.738093358 0.54576983 -0.72576533 -0.23930016 1.1534654
#>  [9,] 1.3436155 -0.373039241 1.53507794 -0.24290707  0.55028323 1.7189182
#> [10,] 1.3391620  0.119827511 1.09724980  0.06461521  0.38346544 1.6564491
#> [11,] 1.5312841  0.151595513 1.48292187  0.45756926  0.50350002 1.8052747
#> [12,] 1.4843813  0.118234501 1.41811320  0.39862832  0.43673350 1.8035451
#> [13,] 1.4205796 -0.276637145 1.54901300 -0.11614056  0.60370986 1.7165007
#> [14,] 0.8998494 -0.302020243 0.76230486 -0.14014666 -0.03892964 1.2317235
#> [15,] 1.5486886 -0.241070244 1.66879549 -0.02859522  0.65104268 1.9193463
#> [16,] 2.0161529  0.182679700 1.97628291  0.45760364  0.79126902 2.2599031
#> [17,] 1.6970836 -0.053634500 1.70631653  0.06737149  0.72161461 1.9632275
#> [18,] 1.0638896 -0.175867710 1.19147525 -0.09511710  0.48396953 1.4972638
#> [19,] 1.4706784  0.321361013 1.35060269  0.37099558  0.65597759 1.8029747
#> [20,] 1.4672325  0.263184239 1.19020008  0.39408495  0.57311792 1.8361600
#> [21,] 2.0659048  0.578778126 1.85255996  0.74417077  0.88366206 2.2552282
#> [22,] 0.9763297 -0.268101952 0.71522106 -0.20762534 -0.04593970 1.4788349
#> [23,] 1.8176848  0.407984226 1.78068510  0.63128043  0.75590905 2.0164048
#> [24,] 1.3136778 -0.124049537 1.30034560 -0.28801250  0.66599838 1.8046956
#> [25,] 1.0084558 -0.720886319 1.12418785 -0.60892042  0.15526126 1.4208587
#> [26,] 1.9089303  0.260376446 1.84289576  0.63391026  0.59644841 2.1043248
#> [27,] 1.5111394 -0.007750357 1.40460712 -0.02041930  0.60824817 1.7692709
#> [28,] 1.4784458  0.125809375 1.40353047  0.37419025  0.47993707 1.7649936
#> [29,] 1.5302394  0.307947674 1.18732471  0.23451800  0.43123807 1.7881823
#> [30,] 2.0176640  0.173688505 1.99461292  0.40347132  0.79531190 2.2273052
#> [31,] 1.5547145 -0.002441988 1.55503830  0.08526831  0.69801575 1.9276342
#> [32,] 1.7734262  0.001017226 1.78855735  0.23897497  0.75332179 2.0051874
#> [33,] 2.0394534  0.254130623 2.01779886  0.46915858  0.93720747 2.2504154
#> [34,] 1.5997614  0.064175467 1.51251388  0.41661191  0.47542962 1.9408492
#> [35,] 1.1187769 -0.172854272 0.86383947 -0.13172704  0.06078221 1.4807214
#> [36,] 1.2661239 -0.298461104 1.14725940 -0.32227268  0.26751002 1.6066616
#> [37,] 1.8541182  0.114828464 1.85278190  0.32596167  0.75638449 2.0849681
#> [38,] 1.0588580 -0.307466115 1.07350514 -0.23115184  0.28362676 1.4235928
#> [39,] 1.5041492  0.369602563 1.40287383  0.60274947  0.70760449 1.8865871
#> [40,] 1.5017081  0.229059857 1.59125369  0.33827813  0.82004022 1.8838336
#>              [,7]        [,8]      [,9]        [,10]
#>  [1,] -0.82076185 -0.08282145 0.6435897 -0.600854547
#>  [2,] -0.45399147  0.56080862 1.5119559 -0.194844326
#>  [3,] -0.23490359  0.83799681 1.4361489 -0.111730666
#>  [4,] -0.19668242  0.86368165 1.6086345 -0.182553899
#>  [5,] -0.19466048  0.97915158 1.6932463 -0.157581484
#>  [6,] -1.30012911 -0.59633435 0.1029850 -0.924316049
#>  [7,] -0.63441153  0.50756818 1.4331612 -0.414532914
#>  [8,] -1.30281916 -0.35290402 0.6405742 -0.843391325
#>  [9,] -0.64875370  0.40133299 1.6056733 -0.309892104
#> [10,] -0.43867371  0.31798949 1.1320152 -0.429242125
#> [11,] -0.36509803  0.65683006 1.3421927 -0.338971781
#> [12,] -0.38633439  0.54891798 1.2888072 -0.395183857
#> [13,] -0.54121016  0.54228640 1.6470316 -0.202193825
#> [14,] -0.87950288  0.02585977 0.6172898 -0.680112696
#> [15,] -0.58551761  0.50445868 1.7337337 -0.282391676
#> [16,] -0.31824709  0.93605008 1.9087463 -0.223639656
#> [17,] -0.42088903  0.73693596 1.7772597 -0.141418053
#> [18,] -0.62584500  0.21019638 1.2248025 -0.369408908
#> [19,] -0.24163797  0.65885460 1.3652353 -0.099978843
#> [20,] -0.25118844  0.51570917 1.2481813 -0.366914821
#> [21,] -0.04125848  0.99719366 1.7366148 -0.089313796
#> [22,] -1.02252439 -0.11121517 0.7185413 -0.927797074
#> [23,] -0.09275145  0.93593641 1.7163271 -0.064356356
#> [24,] -0.67923469  0.27974933 1.5041823 -0.396257776
#> [25,] -1.19212154  0.05688536 1.0952863 -0.642693044
#> [26,] -0.22006767  0.93586629 1.6670077 -0.240709459
#> [27,] -0.43489350  0.59516415 1.4949650 -0.168079704
#> [28,] -0.32410800  0.53263204 1.2511406 -0.405377197
#> [29,] -0.44077808  0.46134140 1.2554536 -0.356095857
#> [30,] -0.33629305  0.93693615 1.9870900 -0.144942822
#> [31,] -0.46918258  0.56998974 1.5968922 -0.215252943
#> [32,] -0.31003184  0.78191854 1.7968840 -0.136749434
#> [33,] -0.10843538  1.06599371 1.9711732  0.005917862
#> [34,] -0.41972344  0.68753483 1.3987882 -0.330016199
#> [35,] -0.84231258  0.21536152 0.8386011 -0.669409029
#> [36,] -0.89746497  0.20412077 0.9712580 -0.517207214
#> [37,] -0.36470591  0.83725473 1.8616325 -0.198345408
#> [38,] -0.71063841  0.14145209 1.1135510 -0.525337110
#> [39,] -0.06581065  0.66863922 1.3915509 -0.231461424
#> [40,] -0.14779926  0.72211697 1.6446887 -0.068566424
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2630087 -0.2008127
#>  [2,] 0.2886683 -0.2336991
#>  [3,] 0.3008498 -0.2290653
#>  [4,] 0.3030179 -0.2291816
#>  [5,] 0.3056587 -0.2322653
#>  [6,] 0.2409375 -0.1762335
#>  [7,] 0.2965030 -0.2338577
#>  [8,] 0.2626491 -0.2114914
#>  [9,] 0.2827117 -0.2415415
#> [10,] 0.2871830 -0.2172976
#> [11,] 0.2965232 -0.2245405
#> [12,] 0.2953488 -0.2244858
#> [13,] 0.2854856 -0.2400920
#> [14,] 0.2735194 -0.2030736
#> [15,] 0.2890956 -0.2439657
#> [16,] 0.3054799 -0.2407063
#> [17,] 0.2944123 -0.2410748
#> [18,] 0.2744274 -0.2234599
#> [19,] 0.2918567 -0.2217409
#> [20,] 0.2929827 -0.2190601
#> [21,] 0.3082775 -0.2295050
#> [22,] 0.2759557 -0.2095380
#> [23,] 0.3031708 -0.2307410
#> [24,] 0.2796054 -0.2326788
#> [25,] 0.2712982 -0.2304673
#> [26,] 0.3076044 -0.2341757
#> [27,] 0.2892679 -0.2313530
#> [28,] 0.2947687 -0.2224522
#> [29,] 0.2929789 -0.2181061
#> [30,] 0.3041833 -0.2423550
#> [31,] 0.2905143 -0.2356392
#> [32,] 0.2976358 -0.2410652
#> [33,] 0.3053491 -0.2413298
#> [34,] 0.2986097 -0.2305493
#> [35,] 0.2820236 -0.2114061
#> [36,] 0.2821737 -0.2221693
#> [37,] 0.3003520 -0.2395427
#> [38,] 0.2752853 -0.2219709
#> [39,] 0.2958713 -0.2218521
#> [40,] 0.2920881 -0.2311248
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3860797 -0.3860797
#>  [2,]  0.3722988 -0.3722988
#>  [3,] -0.6294633  0.6294633
#>  [4,] -0.6299960  0.6299960
#>  [5,]  0.3686707 -0.3686707
#>  [6,] -0.6028061  0.6028061
#>  [7,]  0.3704327 -0.3704327
#>  [8,]  0.3836367 -0.3836367
#>  [9,]  0.3718582 -0.3718582
#> [10,] -0.6235117  0.6235117
#> [11,]  0.3726035 -0.3726035
#> [12,]  0.3728909 -0.3728909
#> [13,]  0.3715489 -0.3715489
#> [14,]  0.3830570 -0.3830570
#> [15,] -0.6301968  0.6301968
#> [16,]  0.3667497 -0.3667497
#> [17,]  0.3692380 -0.3692380
#> [18,] -0.6219627  0.6219627
#> [19,] -0.6256495  0.6256495
#> [20,] -0.6252852  0.6252852
#> [21,] -0.6312964  0.6312964
#> [22,]  0.3809557 -0.3809557
#> [23,] -0.6303950  0.6303950
#> [24,] -0.6253418  0.6253418
#> [25,] -0.6228741  0.6228741
#> [26,]  0.3677736 -0.3677736
#> [27,]  0.3727071 -0.3727071
#> [28,] -0.6264977  0.6264977
#> [29,]  0.3749392 -0.3749392
#> [30,]  0.3666679 -0.3666679
#> [31,]  0.3714145 -0.3714145
#> [32,]  0.3684898 -0.3684898
#> [33,] -0.6333647  0.6333647
#> [34,]  0.3707130 -0.3707130
#> [35,]  0.3790860 -0.3790860
#> [36,] -0.6234794  0.6234794
#> [37,]  0.3682121 -0.3682121
#> [38,] -0.6218143  0.6218143
#> [39,] -0.6266153  0.6266153
#> [40,] -0.6278987  0.6278987
#> 
#> $L
#> [1] 0.6938637 0.6942652 0.7748961 0.6928012 0.7307505 0.6989286
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4637681