Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13          V14
#>  [1,] -0.3872036 -0.05286059 -0.18365631 -0.03942017  0.026062186 -0.003740586
#>  [2,] -0.3216263 -0.05023810 -0.06959853 -0.08090097 -0.157289000 -0.090977843
#>  [3,] -0.3019412 -0.16590713 -0.19842594 -0.11179288 -0.128024271 -0.033465734
#>  [4,] -0.4071404 -0.05094809 -0.06839880 -0.07208661 -0.066291847  0.053631692
#>  [5,] -0.3807879 -0.23307995 -0.06297324 -0.02544152  0.023641432 -0.066422927
#>  [6,] -0.2625233 -0.21968880 -0.14074552 -0.21343605 -0.120909179 -0.062530219
#>  [7,] -0.1892101 -0.16494602 -0.14493124 -0.18474426 -0.127412348 -0.200907301
#>  [8,] -0.1925725 -0.07605144 -0.16780531 -0.16092362 -0.117624111 -0.144137544
#>  [9,] -0.2938210 -0.15473888 -0.08329250 -0.20768127 -0.072729280 -0.103738520
#> [10,] -0.3360535 -0.10283441 -0.10224362 -0.11067119  0.008124487  0.033338308
#>               V15          V16          V17         V18         V19         V2
#>  [1,]  0.11036436  0.062484591  0.083432930  0.09211521  0.03037698 -0.3975172
#>  [2,] -0.06001310 -0.005700148  0.054377844  0.02531934  0.01795726 -0.3377953
#>  [3,] -0.08246087 -0.204998802 -0.209917417 -0.24648239 -0.12050096 -0.2719587
#>  [4,]  0.12813163  0.216742762  0.253811040  0.24821911  0.25002805 -0.4489328
#>  [5,]  0.07947029  0.051891520  0.006423023  0.10237211  0.13489094 -0.2613125
#>  [6,]  0.02652215  0.039780491  0.023597044 -0.07735347 -0.10850219 -0.2342170
#>  [7,] -0.03535577 -0.143493870 -0.183956087 -0.15654527 -0.12080751 -0.3385294
#>  [8,] -0.14841537  0.044488117 -0.068144694  0.04376427  0.03524168 -0.1994739
#>  [9,] -0.12566980 -0.099604607 -0.153697355 -0.10598656 -0.02968829 -0.2111788
#> [10,] -0.11318259 -0.073464082 -0.076329490 -0.01697853  0.06455268 -0.2225000
#>                V20          V21         V22         V23         V24
#>  [1,]  0.173924755  0.105621756  0.12110128  0.11032439  0.15998090
#>  [2,] -0.008387142  0.126025397  0.21955046  0.13356890  0.14880327
#>  [3,] -0.138881704 -0.095498735 -0.02480794  0.04463742 -0.02615632
#>  [4,]  0.239328926  0.199092380  0.17001829  0.25406436  0.22498652
#>  [5,]  0.137036612  0.042489936  0.02094212  0.10547835  0.06930275
#>  [6,] -0.012668650  0.051071489  0.05986770  0.02498387  0.01761482
#>  [7,]  0.086396518 -0.015483717  0.09636565 -0.07439283 -0.06692049
#>  [8,] -0.016275015 -0.001705516  0.07752311 -0.08799316  0.02946600
#>  [9,] -0.133877217 -0.040443544 -0.08567052 -0.00076068 -0.06664127
#> [10,]  0.019045540  0.040099695  0.11481358  0.05060441  0.08152933
#>                 V25          V26         V27         V28         V29         V3
#>  [1,]  0.2073460355  0.203000264 0.223685695  0.07056204  0.16912507 -0.2662017
#>  [2,]  0.2157268974  0.209021771 0.193373739  0.22238995  0.26092090 -0.2460660
#>  [3,]  0.0954024874  0.006880896 0.040548358  0.11086883  0.14626569 -0.3076913
#>  [4,]  0.2978361955  0.358439388 0.287284740  0.27359579  0.11436573 -0.4181439
#>  [5,]  0.0044214303  0.025190472 0.139829122  0.16211238  0.19242887 -0.3468326
#>  [6,]  0.0785641800  0.104874422 0.112830154  0.20194089  0.04240123 -0.3343088
#>  [7,] -0.0009085537  0.126858556 0.150062200  0.04933028 -0.04999684 -0.2383456
#>  [8,]  0.0291151124  0.118709211 0.008748479  0.18846475  0.16346926 -0.1565749
#>  [9,] -0.0601752202 -0.059881107 0.001907901 -0.06416851 -0.11729462 -0.3381602
#> [10,]  0.0343366423  0.068440885 0.095881615  0.18273274  0.07011560 -0.2167438
#>                V30         V31          V32         V33         V34         V35
#>  [1,]  0.078320200 -0.04007485 -0.081072481 -0.05231071 -0.03961116 -0.11455551
#>  [2,]  0.213859317  0.05911027  0.085599204  0.06909033 -0.02336866 -0.01874798
#>  [3,]  0.019974297 -0.06236827 -0.034652338 -0.02994222 -0.01732281 -0.14941624
#>  [4,]  0.045859820  0.04691355  0.011957185 -0.01396959 -0.09843407 -0.04391888
#>  [5,]  0.175095822  0.11785618  0.043314288 -0.02222018  0.03182212  0.21185532
#>  [6,]  0.070747248  0.05790926  0.006823409  0.03684607 -0.04190462  0.13222998
#>  [7,] -0.019321605 -0.03915040 -0.018100579 -0.14957645 -0.15551061 -0.07734312
#>  [8,]  0.040948034  0.09782312 -0.073551510 -0.06200431  0.06016512  0.01098285
#>  [9,] -0.005047595 -0.05365919 -0.048673777 -0.04141033 -0.03792710 -0.08133772
#> [10,]  0.132843264  0.10504892 -0.103027053 -0.06253411  0.05741991 -0.09104473
#>                V36          V37         V38         V39         V4          V40
#>  [1,] -0.072249235 -0.106602723 -0.14160748 -0.08552901 -0.2774359 -0.086029600
#>  [2,] -0.005232832  0.119988688 -0.05668778 -0.04686115 -0.2423375 -0.089231243
#>  [3,] -0.058678037 -0.080263708 -0.09377398 -0.04472630 -0.2698397  0.008531236
#>  [4,]  0.041446142  0.057603164  0.04394608  0.02247827 -0.2700669  0.064319600
#>  [5,]  0.078849205  0.018126902  0.14897585  0.02811584 -0.3844458  0.020256687
#>  [6,]  0.144223896 -0.016197091 -0.10640933  0.03493785 -0.2339571 -0.072518716
#>  [7,]  0.047629593 -0.017884308 -0.17897058 -0.10175854 -0.2000207 -0.197316335
#>  [8,] -0.056552487 -0.135975905 -0.03938926 -0.10055947 -0.2355700 -0.110788328
#>  [9,] -0.014663204 -0.109847937 -0.16107549 -0.19990946 -0.2255229 -0.072039842
#> [10,] -0.106300762  0.003501384 -0.14569400 -0.08869460 -0.1834570  0.012745498
#>               V41         V42          V43         V44         V45         V46
#>  [1,] -0.09066647  0.03184826 -0.071290700 -0.19891615 -0.22174144 -0.16478867
#>  [2,] -0.08319621 -0.08725146 -0.124990787 -0.12950276 -0.03445660 -0.05431750
#>  [3,] -0.02145559 -0.16813133 -0.035547240 -0.17900652 -0.11060635 -0.09925674
#>  [4,] -0.03659561 -0.07095857 -0.125557968 -0.08842501 -0.20347077 -0.26239894
#>  [5,] -0.03063703  0.03870277 -0.054434769  0.00912791 -0.12433296 -0.18471818
#>  [6,]  0.02869429 -0.05047137 -0.001529956 -0.10212481 -0.13877772 -0.05465338
#>  [7,] -0.19633319 -0.10984700 -0.213929258 -0.08399270 -0.10934413 -0.27600498
#>  [8,] -0.02227925 -0.07016047 -0.053319866 -0.12828983 -0.21828317 -0.10847870
#>  [9,] -0.11552211 -0.09773595 -0.075341702 -0.21351198 -0.08040271 -0.24136999
#> [10,] -0.15042682 -0.13642329 -0.047290370 -0.00829675 -0.09588169 -0.07823394
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.2149610 -0.1723639 -0.2457036 -0.3539488 -0.3327662 -0.4263335
#>  [2,] -0.1133745 -0.1990065 -0.2263586 -0.2474535 -0.2265302 -0.3620817
#>  [3,] -0.2234978 -0.2352892 -0.3394249 -0.2756267 -0.2096428 -0.2059066
#>  [4,] -0.1593334 -0.2362181 -0.3111477 -0.2884040 -0.4248207 -0.3222307
#>  [5,] -0.2470924 -0.2585204 -0.3950763 -0.3292636 -0.3118352 -0.4576278
#>  [6,] -0.2679628 -0.2691246 -0.3068632 -0.3009798 -0.3834589 -0.4119734
#>  [7,] -0.2000830 -0.2585439 -0.2425197 -0.2267615 -0.2934099 -0.2064914
#>  [8,] -0.1609719 -0.2299314 -0.1574529 -0.2587483 -0.2615911 -0.2397928
#>  [9,] -0.2610827 -0.2716101 -0.1996869 -0.1370475 -0.3710117 -0.3415393
#> [10,] -0.1231334 -0.2259182 -0.3406124 -0.1689716 -0.3292969 -0.1983619
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3515740 -0.4548220 -0.3430283 -0.4190563 -0.4417202 -0.3216417
#>  [2,] -0.2149181 -0.1864406 -0.2740253 -0.2744379 -0.3486840 -0.2820679
#>  [3,] -0.3422845 -0.3891411 -0.4083340 -0.2284564 -0.3433290 -0.3249488
#>  [4,] -0.4897503 -0.4578115 -0.3435948 -0.4967036 -0.4149428 -0.4083627
#>  [5,] -0.3154877 -0.3999395 -0.3882261 -0.2937785 -0.3320811 -0.3493435
#>  [6,] -0.3007072 -0.3675007 -0.3680963 -0.2594785 -0.4247456 -0.3941036
#>  [7,] -0.2804086 -0.1972579 -0.1992339 -0.2568556 -0.2492484 -0.3483831
#>  [8,] -0.3532692 -0.1703439 -0.2514071 -0.2763846 -0.3103960 -0.2504937
#>  [9,] -0.3919769 -0.3800650 -0.3784747 -0.3597641 -0.3044069 -0.3739637
#> [10,] -0.3269869 -0.2337473 -0.3428837 -0.3016123 -0.3466535 -0.3546405
#>              V58        V59         V6        V60          V7          V8
#>  [1,] -0.4688794 -0.3316020 -0.2686047 -0.4717768 -0.28835422 -0.15466073
#>  [2,] -0.3775203 -0.2031263 -0.1785963 -0.2586911 -0.13416995 -0.18159204
#>  [3,] -0.2123283 -0.2259162 -0.1408100 -0.2921105 -0.10590997 -0.21022314
#>  [4,] -0.3367764 -0.4426542 -0.2601606 -0.4971808 -0.22165110 -0.16104176
#>  [5,] -0.4386573 -0.4230826 -0.1837277 -0.4071073 -0.26404970 -0.19490614
#>  [6,] -0.3640226 -0.4008095 -0.1692353 -0.2694928 -0.23124655 -0.28030582
#>  [7,] -0.2565888 -0.3228472 -0.1911921 -0.1926319 -0.16350115 -0.25051529
#>  [8,] -0.3634935 -0.3035002 -0.2111725 -0.2209660 -0.13106067 -0.26920667
#>  [9,] -0.2540407 -0.3382346 -0.2018538 -0.3748212 -0.12484852 -0.20123085
#> [10,] -0.3044500 -0.3553474 -0.1426629 -0.3311216 -0.09412577 -0.07896197
#>                V9
#>  [1,] -0.09778585
#>  [2,] -0.11973000
#>  [3,] -0.10793371
#>  [4,] -0.20594434
#>  [5,] -0.13717693
#>  [6,] -0.11376287
#>  [7,] -0.11548389
#>  [8,] -0.10553839
#>  [9,] -0.14295652
#> [10,] -0.13930737
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.06999989  0.05992789 -0.02870050  0.05296574 -0.05214372 -0.11585188
#> [2,] -0.05052395 -0.11346757 -0.07458992 -0.15926524 -0.07848537  0.02444531
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.04683401 -0.03712343 -0.06700891 -0.04106575
#> [2,]  0.09531333  0.10706640  0.08896317 -0.01879897
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -3.700411e-05 -0.0005404891 -0.0004698063 -0.0002954038 -0.0002005791
#>  [2,] -7.227086e-05 -0.0008520647 -0.0007552185 -0.0004754589 -0.0003737065
#>  [3,] -2.665125e-05 -0.0004359518 -0.0004313500 -0.0003597744 -0.0002124653
#>  [4,] -5.275650e-05 -0.0006736747 -0.0005550017 -0.0003094918 -0.0002573870
#>  [5,]  4.968811e-05  0.0003691776  0.0003748149  0.0003431640  0.0003274197
#>  [6,]  1.997529e-04  0.0019516677  0.0018840448  0.0015411425  0.0013361125
#>  [7,]  8.085122e-05  0.0009534052  0.0009627450  0.0008166399  0.0005511670
#>  [8,]  1.790390e-04  0.0018893500  0.0018518521  0.0015464639  0.0012494778
#>  [9,]  6.118101e-05  0.0007794830  0.0007765555  0.0006278639  0.0003982291
#> [10,]  9.688793e-05  0.0009076509  0.0009008771  0.0007767681  0.0006640857
#>                 V14           V15           V16           V17           V18
#>  [1,] -1.511348e-04  8.470649e-05  2.399316e-04  2.215742e-04  8.249243e-05
#>  [2,] -3.871258e-04 -1.003677e-04  9.423130e-05  5.427682e-05 -1.641171e-04
#>  [3,] -9.503253e-05  1.095507e-04  2.365823e-04  1.748328e-04  2.304082e-05
#>  [4,] -3.017974e-04 -1.797051e-04 -1.033873e-04 -3.525554e-05 -8.576145e-05
#>  [5,]  3.717256e-04  3.898097e-04  4.275521e-04  4.337245e-04  4.836597e-04
#>  [6,]  1.394100e-03  1.034222e-03  8.806603e-04  1.040313e-03  1.501513e-03
#>  [7,]  4.621448e-04  2.471359e-04  1.047261e-04  2.999659e-04  5.966139e-04
#>  [8,]  1.192114e-03  7.460381e-04  5.271673e-04  7.555108e-04  1.276585e-03
#>  [9,]  3.298960e-04  1.430723e-04  6.518843e-06  1.741414e-04  4.183178e-04
#> [10,]  6.559488e-04  4.968038e-04  4.389897e-04  5.043728e-04  7.266227e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0002301596 -3.293348e-05 -0.0007045140 -0.0007058897 -0.0004688598
#>  [2,] -0.0006381176 -6.822746e-05 -0.0013377418 -0.0013778147 -0.0010113228
#>  [3,] -0.0002123716 -2.308042e-05 -0.0004759705 -0.0004957504 -0.0003848623
#>  [4,] -0.0003419611 -6.225432e-05 -0.0008952298 -0.0009197479 -0.0006089515
#>  [5,]  0.0006286477  6.112657e-05  0.0007683727  0.0008151547  0.0007655936
#>  [6,]  0.0024609374  2.194975e-04  0.0035559419  0.0037597844  0.0033310224
#>  [7,]  0.0010113445  8.062693e-05  0.0014536805  0.0015713139  0.0014057977
#>  [8,]  0.0022226420  1.917616e-04  0.0032535847  0.0034444951  0.0030650941
#>  [9,]  0.0007657247  5.818664e-05  0.0011516659  0.0012312241  0.0010707404
#> [10,]  0.0011683238  1.084293e-04  0.0016478542  0.0017336045  0.0015412458
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0002374129 -2.041656e-05  2.475012e-05 -3.858026e-05 -6.331709e-07
#>  [2,] -0.0006600471 -3.452065e-04 -2.091704e-04 -3.044791e-04 -2.547623e-04
#>  [3,] -0.0002379981 -4.936659e-05  5.613014e-05  4.731429e-05  2.256398e-05
#>  [4,] -0.0003285356 -1.678077e-04 -1.756693e-04 -3.042083e-04 -2.685586e-04
#>  [5,]  0.0007269999  7.274824e-04  6.442007e-04  7.025712e-04  7.663293e-04
#>  [6,]  0.0028989611  2.475733e-03  1.997114e-03  2.223291e-03  2.379466e-03
#>  [7,]  0.0011125601  7.506547e-04  3.599042e-04  3.833254e-04  5.212463e-04
#>  [8,]  0.0026086796  2.101736e-03  1.574689e-03  1.732487e-03  1.921433e-03
#>  [9,]  0.0007634725  4.107909e-04  6.989915e-05  6.559499e-05  1.686399e-04
#> [10,]  0.0013527529  1.183665e-03  9.722241e-04  1.074922e-03  1.177897e-03
#>                 V28           V29            V3           V30           V31
#>  [1,] -1.407736e-04 -0.0000459619 -3.295813e-05 -1.422973e-04  0.0004930699
#>  [2,] -4.856493e-04 -0.0003700821 -6.986603e-05 -5.279901e-04  0.0004240790
#>  [3,] -8.914981e-05 -0.0000398940 -2.689814e-05 -1.142961e-05  0.0004745031
#>  [4,] -4.100773e-04 -0.0002321682 -6.479743e-05 -3.803887e-04  0.0002092915
#>  [5,]  8.557549e-04  0.0007912856  6.110175e-05  7.556711e-04  0.0005253995
#>  [6,]  2.845737e-03  0.0025028593  2.211591e-04  2.472416e-03  0.0007004636
#>  [7,]  7.673951e-04  0.0006457865  8.763092e-05  5.560665e-04 -0.0004751747
#>  [8,]  2.388621e-03  0.0020730308  1.957913e-04  1.971182e-03  0.0001488184
#>  [9,]  3.658811e-04  0.0002686610  6.590556e-05  2.318888e-04 -0.0006276863
#> [10,]  1.414656e-03  0.0012706679  1.091791e-04  1.214601e-03  0.0004249738
#>                 V32           V33           V34           V35           V36
#>  [1,]  2.908249e-04  1.157918e-04  2.094613e-04  4.944316e-04  7.337089e-04
#>  [2,]  2.141969e-04 -2.693227e-05  9.330544e-05  4.600088e-04  8.086051e-04
#>  [3,]  2.894687e-04  1.927468e-04  3.322817e-04  5.779462e-04  7.405668e-04
#>  [4,]  4.394559e-05 -6.572701e-05  2.543116e-05  2.420252e-04  4.209620e-04
#>  [5,]  4.221774e-04  3.965884e-04  3.573818e-04  3.232223e-04  2.997149e-04
#>  [6,]  6.898536e-04  9.200274e-04  6.662175e-04  1.527205e-04 -2.757280e-04
#>  [7,] -3.552473e-04 -1.718902e-04 -3.302060e-04 -6.497611e-04 -8.514675e-04
#>  [8,]  2.527825e-04  5.151517e-04  2.053319e-04 -3.873917e-04 -8.250135e-04
#>  [9,] -4.796884e-04 -3.026790e-04 -4.584903e-04 -7.794149e-04 -9.917293e-04
#> [10,]  4.066217e-04  4.841774e-04  3.422566e-04  9.649176e-05 -8.941339e-05
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0004537981 -1.270855e-04 -0.0002242272 -9.875469e-05  5.765848e-05
#>  [2,]  0.0004633590 -3.543065e-04 -0.0004855231 -1.688826e-04 -7.783183e-05
#>  [3,]  0.0004481252 -3.586844e-05 -0.0001508427 -7.651880e-05  3.188584e-05
#>  [4,]  0.0001408784 -3.529825e-04 -0.0003685448 -1.273263e-04 -7.171714e-05
#>  [5,]  0.0003485466  4.841725e-04  0.0004983801  7.986035e-05  4.179085e-04
#>  [6,]  0.0002799740  1.624983e-03  0.0018747442  3.927291e-04  1.254404e-03
#>  [7,] -0.0003500118  5.000024e-04  0.0006493251  1.757894e-04  3.263356e-04
#>  [8,] -0.0001217339  1.353942e-03  0.0016348320  3.731681e-04  1.005781e-03
#>  [9,] -0.0005337807  2.585961e-04  0.0004276708  1.436309e-04  1.518048e-04
#> [10,]  0.0001892599  8.043025e-04  0.0009134491  1.870326e-04  6.211815e-04
#>                 V41           V42           V43           V44           V45
#>  [1,] -9.211803e-06 -8.719623e-05 -0.0002119970 -0.0003453006 -0.0005909156
#>  [2,] -1.521966e-04 -2.587861e-04 -0.0004122049 -0.0005685660 -0.0008902762
#>  [3,] -1.472381e-05 -7.710099e-05 -0.0001796771 -0.0003118042 -0.0004630492
#>  [4,] -1.350876e-04 -1.926006e-04 -0.0002761959 -0.0003864231 -0.0006748587
#>  [5,]  4.207154e-04  4.182820e-04  0.0003890632  0.0003787096  0.0004303669
#>  [6,]  1.339587e-03  1.433351e-03  0.0015468608  0.0017248552  0.0021656969
#>  [7,]  3.630205e-04  4.585829e-04  0.0005940028  0.0007741712  0.0009914478
#>  [8,]  1.115590e-03  1.245566e-03  0.0014165958  0.0016581837  0.0021071401
#>  [9,]  1.840800e-04  2.675922e-04  0.0003989939  0.0005627911  0.0007850419
#> [10,]  6.571084e-04  7.078234e-04  0.0007558836  0.0008365066  0.0010373165
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0004970017 -0.0002904088 -0.0002332989 -0.0001372890 -6.048824e-05
#>  [2,] -0.0007370682 -0.0004472231 -0.0003573221 -0.0002130603 -1.230915e-04
#>  [3,] -0.0003825890 -0.0002528941 -0.0001881282 -0.0001141659 -6.718143e-05
#>  [4,] -0.0005929169 -0.0003515208 -0.0002881259 -0.0001595016 -8.295276e-05
#>  [5,]  0.0003714682  0.0002713342  0.0001930556  0.0001146081  1.142080e-04
#>  [6,]  0.0018021138  0.0012333092  0.0009140663  0.0005554217  4.655062e-04
#>  [7,]  0.0008047757  0.0005790428  0.0004155362  0.0002596959  2.029046e-04
#>  [8,]  0.0017442137  0.0012034697  0.0008840910  0.0005425162  4.334494e-04
#>  [9,]  0.0006418334  0.0004606633  0.0003323531  0.0002092622  1.523100e-04
#> [10,]  0.0008674104  0.0006001272  0.0004400133  0.0002658220  2.247723e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -4.149362e-05 -3.857431e-05 -2.690340e-05 -1.176457e-05 -1.003964e-05
#>  [2,] -6.624237e-05 -6.218903e-05 -4.299865e-05 -2.246792e-05 -2.135306e-05
#>  [3,] -2.776311e-05 -3.018136e-05 -2.154099e-05 -5.733716e-06 -3.845210e-06
#>  [4,] -5.076728e-05 -4.720209e-05 -3.709805e-05 -1.404670e-05 -1.596425e-05
#>  [5,]  3.622453e-05  2.875535e-05  2.219487e-05  1.142007e-05  1.317536e-05
#>  [6,]  1.671731e-04  1.445574e-04  1.049917e-04  5.176676e-05  5.424669e-05
#>  [7,]  6.337900e-05  6.368949e-05  4.475094e-05  1.141793e-05  1.186994e-05
#>  [8,]  1.552416e-04  1.383390e-04  9.977999e-05  4.374729e-05  4.420899e-05
#>  [9,]  4.965732e-05  4.992086e-05  3.489249e-05  6.595145e-06  7.382218e-06
#> [10,]  7.957511e-05  6.806345e-05  5.066045e-05  2.457126e-05  2.549771e-05
#>                 V55           V56           V57           V58           V59
#>  [1,] -1.320670e-05 -1.857917e-05 -8.567200e-06 -1.881119e-05 -9.090887e-06
#>  [2,] -2.526429e-05 -3.153115e-05 -1.717674e-05 -3.339321e-05 -1.879812e-05
#>  [3,] -5.324058e-06 -1.010209e-05 -2.880494e-06 -1.207315e-05 -5.836907e-06
#>  [4,] -1.590658e-05 -2.212537e-05 -1.202521e-05 -1.921180e-05 -1.209593e-05
#>  [5,]  1.220799e-05  1.315935e-05  1.133623e-05  1.272338e-05  1.060531e-05
#>  [6,]  5.675781e-05  6.819631e-05  4.785524e-05  6.990116e-05  4.680049e-05
#>  [7,]  1.392943e-05  2.396198e-05  1.312199e-05  2.650498e-05  1.553694e-05
#>  [8,]  4.725904e-05  6.055204e-05  3.958866e-05  6.381753e-05  4.048638e-05
#>  [9,]  1.030723e-05  1.908285e-05  8.944730e-06  2.100725e-05  1.138598e-05
#> [10,]  2.561922e-05  3.035392e-05  2.120526e-05  3.141818e-05  2.176152e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -1.668799e-05 -8.307045e-06 -0.0001340501 -0.0002500951 -0.0005134221
#>  [2,] -7.807843e-05 -1.656761e-05 -0.0002667289 -0.0004576419 -0.0008216093
#>  [3,] -4.029989e-05 -5.978403e-06 -0.0001071655 -0.0001867309 -0.0003810502
#>  [4,] -6.631651e-05 -1.158564e-05 -0.0001820132 -0.0003578004 -0.0006669915
#>  [5,]  1.436318e-04  9.916692e-06  0.0001775460  0.0002470772  0.0003512376
#>  [6,]  4.915663e-04  4.356436e-05  0.0007921081  0.0011697043  0.0018451709
#>  [7,]  1.884826e-04  1.873307e-05  0.0003241022  0.0005259501  0.0008670603
#>  [8,]  4.281682e-04  3.939034e-05  0.0007133314  0.0010745058  0.0017550596
#>  [9,]  1.279418e-04  1.470742e-05  0.0002442255  0.0004231542  0.0007108898
#> [10,]  2.379683e-04  2.027961e-05  0.0003600657  0.0005260668  0.0008454269
#> 
#> $vW[[2]]
#>             [,1]        [,2]          [,3]        [,4]        [,5]        [,6]
#> [1,] -0.02768557 -0.03128583 -0.0003486827 -0.03929932 -0.02715913 -0.01106424
#> [2,]  0.02768557  0.03128583  0.0003486827  0.03929932  0.02715913  0.01106424
#>              [,7]         [,8]         [,9]       [,10]
#> [1,]  0.001747775 -0.008593724  0.003994598 -0.02087807
#> [2,] -0.001747775  0.008593724 -0.003994598  0.02087807
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.66427454  0.46726159  0.22820787  0.77887356  0.62399396  0.46193317
#>  [7] -0.10939873  0.19150604 -0.01644154  0.34241497
#> 
#> $B[[2]]
#> [1]  0.07746113 -0.04825843
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0001324110 -0.0003313673  0.0001976768 -0.0002559140  0.0011127579
#>  [6]  0.0032384399  0.0005779451  0.0024679234  0.0001153346  0.0016090809
#> 
#> $vB[[2]]
#> [1] -0.04631302  0.04631302
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#>  [2,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [3,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#>  [4,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#>  [5,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [6,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#>  [7,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#>  [8,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#>  [9,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [10,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [11,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [12,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [13,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [14,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [15,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [16,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [17,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [18,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [19,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [20,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [21,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [22,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [23,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [24,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [25,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [26,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [27,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [28,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [29,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [30,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [31,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [32,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [33,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [34,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [35,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [36,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [37,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [38,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [39,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [40,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#>  [2,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [3,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#>  [4,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#>  [5,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [6,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#>  [7,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#>  [8,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#>  [9,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [10,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [11,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [12,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [13,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [14,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [15,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [16,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [17,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [18,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [19,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [20,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [21,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [22,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [23,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [24,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [25,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [26,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [27,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [28,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [29,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [30,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [31,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [32,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [33,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [34,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [35,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [36,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [37,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [38,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [39,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [40,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#>  [2,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [3,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#>  [4,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#>  [5,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [6,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#>  [7,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#>  [8,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#>  [9,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [10,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [11,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [12,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [13,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [14,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [15,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [16,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [17,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [18,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [19,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [20,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [21,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [22,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [23,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [24,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [25,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [26,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [27,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [28,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [29,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [30,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [31,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [32,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [33,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [34,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [35,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [36,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [37,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [38,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [39,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [40,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#>  [2,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [3,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#>  [4,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#>  [5,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [6,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#>  [7,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#>  [8,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#>  [9,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [10,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [11,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [12,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [13,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [14,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [15,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [16,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [17,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [18,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [19,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [20,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [21,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [22,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [23,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [24,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [25,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [26,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [27,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [28,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [29,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [30,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [31,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [32,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [33,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [34,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [35,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [36,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [37,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [38,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [39,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [40,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#>  [2,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [3,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#>  [4,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#>  [5,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [6,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#>  [7,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#>  [8,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#>  [9,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [10,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [11,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [12,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [13,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [14,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [15,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [16,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [17,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [18,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [19,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [20,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [21,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [22,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [23,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [24,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [25,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [26,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [27,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [28,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [29,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [30,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [31,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [32,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [33,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [34,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [35,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [36,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [37,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [38,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [39,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [40,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#>  [2,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [3,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#>  [4,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#>  [5,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [6,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#>  [7,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#>  [8,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#>  [9,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [10,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [11,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [12,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [13,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [14,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [15,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [16,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [17,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [18,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [19,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [20,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [21,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [22,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [23,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [24,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [25,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [26,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [27,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [28,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [29,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [30,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [31,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [32,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [33,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [34,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [35,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [36,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [37,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [38,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [39,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [40,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7017754 0.8128601 0.3718347 0.9071811 0.8180090 0.6681294 0.2731638
#>  [2,] 0.4909165 0.6556801 0.1891787 0.7866926 0.7230874 0.5118926 0.1359972
#>  [3,] 0.8161644 0.8286930 0.3708497 0.9344513 0.7948072 0.6604600 0.2607207
#>  [4,] 0.8223797 0.8611249 0.4358059 0.9430042 0.8163756 0.6862272 0.3327226
#>  [5,] 0.7795545 0.7851103 0.2800008 0.9295275 0.7994094 0.6112087 0.2443393
#>  [6,] 0.7442455 0.7681267 0.2092340 0.9453531 0.7917748 0.6063505 0.1813221
#>  [7,] 0.8033873 0.8355312 0.4146197 0.9370550 0.8189868 0.6858054 0.3589886
#>  [8,] 0.8240018 0.8478602 0.3792940 0.9418840 0.7841280 0.6284967 0.2985094
#>  [9,] 0.8230371 0.8252583 0.2488375 0.9520896 0.8133507 0.5997168 0.1945637
#> [10,] 0.8195956 0.8462548 0.3739287 0.9450044 0.8315182 0.6902705 0.2970158
#> [11,] 0.5130511 0.6240579 0.2075557 0.7770685 0.7145497 0.5033357 0.1460639
#> [12,] 0.7506391 0.7666722 0.2672433 0.9316048 0.8046719 0.6107761 0.2092525
#> [13,] 0.7749051 0.7788299 0.3468952 0.9257887 0.7596697 0.5958509 0.2694174
#> [14,] 0.7876890 0.7758683 0.2626820 0.9446416 0.7928429 0.5888845 0.2091365
#> [15,] 0.5390578 0.6916069 0.3505373 0.7794431 0.8025863 0.6396146 0.1979928
#> [16,] 0.5472233 0.6737440 0.2189200 0.8429978 0.8097258 0.5557906 0.1259827
#> [17,] 0.6462031 0.8088232 0.3298121 0.8725239 0.7861606 0.6316355 0.1887640
#> [18,] 0.7789842 0.7893618 0.2609931 0.9469626 0.8147702 0.6139431 0.2172355
#> [19,] 0.7135739 0.7752635 0.3418505 0.9058963 0.7896598 0.6664129 0.2533741
#> [20,] 0.7831657 0.7810913 0.2826159 0.9349469 0.7828130 0.6160219 0.2693809
#> [21,] 0.8268619 0.8288346 0.3054131 0.9511530 0.8066385 0.6318292 0.2536892
#> [22,] 0.7794826 0.7903543 0.3570454 0.9234220 0.7889555 0.6280872 0.3157452
#> [23,] 0.7600378 0.8583506 0.4142502 0.9285193 0.8385697 0.7124799 0.3022779
#> [24,] 0.6301964 0.7347653 0.2510373 0.8379885 0.7245339 0.5402783 0.1512785
#> [25,] 0.8357328 0.8446235 0.3148439 0.9599619 0.7995508 0.6322281 0.2514198
#> [26,] 0.8329265 0.8721072 0.4720667 0.9405089 0.8081189 0.6998361 0.3756070
#> [27,] 0.6913642 0.7122304 0.2426779 0.9119936 0.7607464 0.5674650 0.1781789
#> [28,] 0.7940278 0.7515180 0.2169962 0.9513071 0.7419150 0.5308986 0.1720606
#> [29,] 0.7408733 0.8408473 0.3457361 0.9311991 0.8333779 0.6889522 0.2599951
#> [30,] 0.6588895 0.7339220 0.1997769 0.8980086 0.7079913 0.4998706 0.1293169
#> [31,] 0.6013098 0.7339492 0.3552902 0.7941191 0.7892916 0.6443714 0.2374488
#> [32,] 0.6747810 0.7684972 0.2578061 0.8952667 0.7877511 0.5686053 0.1324933
#> [33,] 0.7254304 0.7634008 0.1991542 0.9446798 0.8000223 0.6084639 0.1666365
#> [34,] 0.5718335 0.7649363 0.3098498 0.8283548 0.7650441 0.5916613 0.1597482
#> [35,] 0.8598733 0.8631227 0.3049383 0.9722476 0.8290105 0.6803801 0.2830336
#> [36,] 0.6925702 0.8208490 0.4110452 0.8640023 0.8108326 0.6640155 0.2711506
#> [37,] 0.7760352 0.7921975 0.4758892 0.8915744 0.7563106 0.6775778 0.3618490
#> [38,] 0.8242881 0.8571605 0.4110892 0.9388026 0.7895950 0.6509121 0.3232251
#> [39,] 0.8478894 0.8590863 0.3213880 0.9656349 0.8228841 0.6510738 0.2590680
#> [40,] 0.8103617 0.8095737 0.2958760 0.9458919 0.7788517 0.5933659 0.2393414
#>            [,8]       [,9]     [,10]
#>  [1,] 0.4988346 0.17785009 0.5990872
#>  [2,] 0.3181934 0.09798915 0.4135015
#>  [3,] 0.5190553 0.20382276 0.6227009
#>  [4,] 0.5695218 0.22696077 0.6716232
#>  [5,] 0.5059436 0.20856396 0.5997447
#>  [6,] 0.4163954 0.13883541 0.5199548
#>  [7,] 0.5628707 0.23750508 0.6569548
#>  [8,] 0.5404100 0.20147409 0.6683686
#>  [9,] 0.5032007 0.15032335 0.6080054
#> [10,] 0.5431817 0.20846949 0.6435549
#> [11,] 0.3327997 0.12018093 0.4118920
#> [12,] 0.4598033 0.17182753 0.5577128
#> [13,] 0.4878727 0.20809441 0.5859513
#> [14,] 0.4583329 0.16812909 0.5797431
#> [15,] 0.4486505 0.20119506 0.4955881
#> [16,] 0.3831382 0.12271348 0.4329398
#> [17,] 0.4446590 0.14474935 0.5439759
#> [18,] 0.4673585 0.16958823 0.5748718
#> [19,] 0.4631661 0.19651747 0.5703648
#> [20,] 0.4662770 0.21057174 0.5764094
#> [21,] 0.5235367 0.18965380 0.6399398
#> [22,] 0.5160281 0.22358752 0.6274764
#> [23,] 0.5475175 0.21813313 0.6345142
#> [24,] 0.4079067 0.11897020 0.5022556
#> [25,] 0.5037119 0.17063887 0.6299628
#> [26,] 0.5933809 0.27035913 0.6954812
#> [27,] 0.3899732 0.16747843 0.4783623
#> [28,] 0.3916899 0.13137951 0.5613327
#> [29,] 0.4836773 0.17304356 0.5982376
#> [30,] 0.3636417 0.10251622 0.5001086
#> [31,] 0.4843291 0.23564066 0.5299815
#> [32,] 0.4182785 0.13137346 0.5274141
#> [33,] 0.4080782 0.13153738 0.5068207
#> [34,] 0.4055070 0.12817977 0.5054670
#> [35,] 0.5436108 0.19294004 0.6499741
#> [36,] 0.5335644 0.23321680 0.6243581
#> [37,] 0.5579698 0.29356375 0.6239072
#> [38,] 0.5567199 0.21675041 0.6782758
#> [39,] 0.5202329 0.16965510 0.6454470
#> [40,] 0.4692900 0.17054055 0.6112457
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4981193 0.5018807
#>  [2,] 0.5092155 0.4907845
#>  [3,] 0.5000715 0.4999285
#>  [4,] 0.4947230 0.5052770
#>  [5,] 0.5003333 0.4996667
#>  [6,] 0.5089204 0.4910796
#>  [7,] 0.4924683 0.5075317
#>  [8,] 0.4996310 0.5003690
#>  [9,] 0.5067828 0.4932172
#> [10,] 0.4967969 0.5032031
#> [11,] 0.5072603 0.4927397
#> [12,] 0.5043670 0.4956330
#> [13,] 0.5017140 0.4982860
#> [14,] 0.5064282 0.4935718
#> [15,] 0.4942906 0.5057094
#> [16,] 0.5069634 0.4930366
#> [17,] 0.5043921 0.4956079
#> [18,] 0.5049272 0.4950728
#> [19,] 0.4990502 0.5009498
#> [20,] 0.5014436 0.4985564
#> [21,] 0.5014594 0.4985406
#> [22,] 0.4966501 0.5033499
#> [23,] 0.4945456 0.5054544
#> [24,] 0.5085794 0.4914206
#> [25,] 0.5042171 0.4957829
#> [26,] 0.4905950 0.5094050
#> [27,] 0.5083510 0.4916490
#> [28,] 0.5138990 0.4861010
#> [29,] 0.5003455 0.4996545
#> [30,] 0.5153921 0.4846079
#> [31,] 0.4922900 0.5077100
#> [32,] 0.5100084 0.4899916
#> [33,] 0.5095278 0.4904722
#> [34,] 0.5058492 0.4941508
#> [35,] 0.4994092 0.5005908
#> [36,] 0.4936246 0.5063754
#> [37,] 0.4893221 0.5106779
#> [38,] 0.4969628 0.5030372
#> [39,] 0.5030078 0.4969922
#> [40,] 0.5059816 0.4940184
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]        [,3]     [,4]      [,5]          [,6]
#>  [1,]  0.85576661 1.4687023 -0.52435402 2.279692 1.5029161  0.6997367373
#>  [2,] -0.03633819 0.6441018 -1.45535571 1.305103 0.9598280  0.0475795296
#>  [3,]  1.49057409 1.5763923 -0.52857314 2.657166 1.3541496  0.6653449661
#>  [4,]  1.53255382 1.8246642 -0.25820119 2.806094 1.4919820  0.7825394468
#>  [5,]  1.26307214 1.2956996 -0.94445744 2.579455 1.3826070  0.4523956643
#>  [6,]  1.06815311 1.1977636 -1.32954871 2.850666 1.3356567  0.4319971748
#>  [7,]  1.40760090 1.6253464 -0.34490016 2.700480 1.5094982  0.7805811530
#>  [8,]  1.54369917 1.7179160 -0.49254611 2.785440 1.2898867  0.5257728827
#>  [9,]  1.53706135 1.5523879 -1.10482186 2.989326 1.4719309  0.4042851872
#> [10,]  1.51361001 1.7055238 -0.51539902 2.843936 1.5964252  0.8013843488
#> [11,]  0.05221640 0.5068079 -1.33972240 1.248664 0.9175849  0.0133428300
#> [12,]  1.10202389 1.1896150 -1.00865435 2.611606 1.4157541  0.4505756808
#> [13,]  1.23621842 1.2588609 -0.63271477 2.523730 1.1508696  0.3882065232
#> [14,]  1.31105144 1.2417489 -1.03207509 2.836977 1.3421479  0.3593557960
#> [15,]  0.15655012 0.8076426 -0.61667814 1.262424 1.4025377  0.5736917064
#> [16,]  0.18945768 0.7251678 -1.27197116 1.680704 1.4482293  0.2240954053
#> [17,]  0.60239092 1.4423819 -0.70903521 1.923461 1.3019358  0.5392396191
#> [18,]  1.25975671 1.3210829 -1.04081307 2.882261 1.4813092  0.4639173523
#> [19,]  0.91280568 1.2382742 -0.65505847 2.264528 1.3228759  0.6920053150
#> [20,]  1.28421077 1.2720375 -0.93152263 2.665286 1.2821352  0.4726970974
#> [21,]  1.56354845 1.5773905 -0.82165180 2.968983 1.4283143  0.5400722179
#> [22,]  1.26265399 1.3270626 -0.58821114 2.489776 1.3186412  0.5240195861
#> [23,]  1.15288650 1.8016574 -0.34642248 2.564163 1.6476246  0.9074588528
#> [24,]  0.53305936 1.0189361 -1.09308750 1.643337 0.9670642  0.1614630571
#> [25,]  1.62681447 1.6930396 -0.77756973 3.177063 1.3834892  0.5417873768
#> [26,]  1.60651170 1.9197201 -0.11184950 2.760595 1.4378332  0.8465173817
#> [27,]  0.80650499 0.9062412 -1.13805345 2.338223 1.1567762  0.2715159274
#> [28,]  1.34937699 1.1067249 -1.28325796 2.972304 1.0559455  0.1237520789
#> [29,]  1.05051264 1.6645458 -0.63783521 2.605257 1.6097590  0.7952251268
#> [30,]  0.65834922 1.0146132 -1.38768935 2.175291 0.8856482 -0.0005177991
#> [31,]  0.41092562 1.0147527 -0.59586534 1.349936 1.3206607  0.5943885478
#> [32,]  0.72988950 1.1998450 -1.05740268 2.145704 1.3114223  0.2761631435
#> [33,]  0.97156057 1.1714157 -1.39158900 2.837708 1.3864339  0.4408599606
#> [34,]  0.28933570 1.1799363 -0.80082167 1.574012 1.1805354  0.3708373528
#> [35,]  1.81423839 1.8414721 -0.82389128 3.556288 1.5786309  0.7555191431
#> [36,]  0.81216284 1.5221102 -0.35964634 1.848938 1.4554293  0.6812404273
#> [37,]  1.24270880 1.3382225 -0.09651791 2.106925 1.1325576  0.7426627220
#> [38,]  1.54567462 1.7919036 -0.35946465 2.730501 1.3224861  0.6230508452
#> [39,]  1.71814224 1.8077217 -0.74740030 3.335745 1.5360108  0.6237626299
#> [40,]  1.45236216 1.4472425 -0.86701402 2.861144 1.2589874  0.3778974199
#>             [,7]         [,8]       [,9]         [,10]
#>  [1,] -0.9786296 -0.004661476 -1.5309817  0.4016631366
#>  [2,] -1.8489421 -0.762086873 -2.2197698 -0.3495090226
#>  [3,] -1.0422263  0.076257935 -1.3625710  0.5010280903
#>  [4,] -0.6958969  0.279900327 -1.2255526  0.7155358935
#>  [5,] -1.1290349  0.023775708 -1.3336033  0.4044014317
#>  [6,] -1.5074155 -0.337588428 -1.8249965  0.0798617823
#>  [7,] -0.5797566  0.252820890 -1.1664068  0.6497529866
#>  [8,] -0.8544059  0.161993152 -1.3771066  0.7008156155
#>  [9,] -1.4206243  0.012802780 -1.7320672  0.4389358830
#> [10,] -0.8615492  0.173158223 -1.3341757  0.5908270419
#> [11,] -1.7658123 -0.695549667 -1.9907179 -0.3561493103
#> [12,] -1.3294371 -0.161134647 -1.5727302  0.2318846794
#> [13,] -0.9975807 -0.048518853 -1.3364503  0.3472529190
#> [14,] -1.3301382 -0.167055813 -1.5989452  0.3217188181
#> [15,] -1.3988872 -0.206124734 -1.3788419 -0.0176481800
#> [16,] -1.9369556 -0.476249116 -1.9669814 -0.2698669301
#> [17,] -1.4580616 -0.222274736 -1.7763910  0.1763591610
#> [18,] -1.2818502 -0.130751832 -1.5885483  0.3017564397
#> [19,] -1.0806969 -0.147603126 -1.4082041  0.2833395216
#> [20,] -0.9977659 -0.135097283 -1.3214825  0.3080507669
#> [21,] -1.0790324  0.094216351 -1.4522613  0.5751027996
#> [22,] -0.7733949  0.064134231 -1.2448810  0.5214057245
#> [23,] -0.8364741  0.190645261 -1.2765789  0.5516323735
#> [24,] -1.7246083 -0.372625818 -2.0022184  0.0090223747
#> [25,] -1.0910541  0.014848053 -1.5811062  0.5320572119
#> [26,] -0.5082367  0.377959483 -0.9928013  0.8258711916
#> [27,] -1.5287348 -0.447424721 -1.6036045 -0.0866050209
#> [28,] -1.5710933 -0.440214534 -1.8888161  0.2465724869
#> [29,] -1.0459940 -0.065314043 -1.5642087  0.3981273326
#> [30,] -1.9070117 -0.559592889 -2.1695740  0.0004344869
#> [31,] -1.1667174 -0.062704252 -1.1767300  0.1200699550
#> [32,] -1.8790914 -0.329844302 -1.8888692  0.1097664935
#> [33,] -1.6096552 -0.371915572 -1.8874335  0.0272843664
#> [34,] -1.6601030 -0.382570966 -1.9171495  0.0218689037
#> [35,] -0.9294636  0.174887417 -1.4310185  0.6189251981
#> [36,] -0.9887927  0.134459843 -1.1902356  0.5080877545
#> [37,] -0.5673479  0.232926511 -0.8781381  0.5061655911
#> [38,] -0.7389899  0.227860556 -1.2847049  0.7458593982
#> [39,] -1.0508182  0.080975787 -1.5880736  0.5990852966
#> [40,] -1.1562935 -0.122994672 -1.5818012  0.4525513578
#> 
#> $pre[[3]]
#>              [,1]       [,2]
#>  [1,] -0.14738323 -0.1398605
#>  [2,] -0.09882067 -0.1356869
#>  [3,] -0.14233761 -0.1426235
#>  [4,] -0.15793838 -0.1368298
#>  [5,] -0.13421618 -0.1355495
#>  [6,] -0.11643119 -0.1521168
#>  [7,] -0.15973048 -0.1296013
#>  [8,] -0.14205801 -0.1405822
#>  [9,] -0.12413190 -0.1512647
#> [10,] -0.15265334 -0.1398408
#> [11,] -0.10011240 -0.1291555
#> [12,] -0.12732139 -0.1447897
#> [13,] -0.13090702 -0.1377629
#> [14,] -0.12252225 -0.1482366
#> [15,] -0.14340294 -0.1205644
#> [16,] -0.11373684 -0.1415924
#> [17,] -0.13016430 -0.1477333
#> [18,] -0.12791054 -0.1476198
#> [19,] -0.14151558 -0.1377165
#> [20,] -0.13147175 -0.1372460
#> [21,] -0.13711916 -0.1429566
#> [22,] -0.14404864 -0.1306487
#> [23,] -0.16008598 -0.1382677
#> [24,] -0.10708549 -0.1414064
#> [25,] -0.13310882 -0.1499774
#> [26,] -0.16657424 -0.1289497
#> [27,] -0.11218026 -0.1455874
#> [28,] -0.10196066 -0.1575710
#> [29,] -0.14606020 -0.1474420
#> [30,] -0.09296224 -0.1545500
#> [31,] -0.14675401 -0.1159115
#> [32,] -0.11422227 -0.1542614
#> [33,] -0.11568556 -0.1538013
#> [34,] -0.12131634 -0.1447141
#> [35,] -0.14544359 -0.1430806
#> [36,] -0.15473308 -0.1292302
#> [37,] -0.15990613 -0.1171879
#> [38,] -0.14945998 -0.1373110
#> [39,] -0.13853224 -0.1505634
#> [40,] -0.12496998 -0.1488976
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5018807 -0.5018807
#>  [2,]  0.4907845 -0.4907845
#>  [3,] -0.5000715  0.5000715
#>  [4,] -0.4947230  0.4947230
#>  [5,]  0.4996667 -0.4996667
#>  [6,] -0.5089204  0.5089204
#>  [7,] -0.4924683  0.4924683
#>  [8,]  0.5003690 -0.5003690
#>  [9,] -0.5067828  0.5067828
#> [10,] -0.4967969  0.4967969
#> [11,]  0.4927397 -0.4927397
#> [12,]  0.4956330 -0.4956330
#> [13,]  0.4982860 -0.4982860
#> [14,]  0.4935718 -0.4935718
#> [15,] -0.4942906  0.4942906
#> [16,] -0.5069634  0.5069634
#> [17,]  0.4956079 -0.4956079
#> [18,]  0.4950728 -0.4950728
#> [19,] -0.4990502  0.4990502
#> [20,]  0.4985564 -0.4985564
#> [21,]  0.4985406 -0.4985406
#> [22,] -0.4966501  0.4966501
#> [23,] -0.4945456  0.4945456
#> [24,]  0.4914206 -0.4914206
#> [25,]  0.4957829 -0.4957829
#> [26,] -0.4905950  0.4905950
#> [27,] -0.5083510  0.5083510
#> [28,]  0.4861010 -0.4861010
#> [29,]  0.4996545 -0.4996545
#> [30,]  0.4846079 -0.4846079
#> [31,] -0.4922900  0.4922900
#> [32,]  0.4899916 -0.4899916
#> [33,] -0.5095278  0.5095278
#> [34,]  0.4941508 -0.4941508
#> [35,] -0.4994092  0.4994092
#> [36,] -0.4936246  0.4936246
#> [37,] -0.4893221  0.4893221
#> [38,]  0.5030372 -0.5030372
#> [39,]  0.4969922 -0.4969922
#> [40,]  0.4940184 -0.4940184
#> 
#> $L
#> [1] 0.7030214 0.7692483 0.7742138 0.7189300 0.6916975 0.6867774
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4637681