Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1           V10           V11         V12         V13
#>  [1,] -0.3137286 -1.057591e-01  0.0025944223 -0.14608456 -0.10517340
#>  [2,] -0.3751159 -1.495031e-01 -0.1762682530 -0.02774413 -0.05883023
#>  [3,] -0.3625239 -7.360142e-02 -0.1802496907 -0.05688008 -0.15454784
#>  [4,] -0.3398182 -2.330658e-05 -0.0835161537 -0.13156670 -0.17901120
#>  [5,] -0.2176599 -5.160278e-02 -0.1000188502  0.03088082  0.04167206
#>  [6,] -0.3217802 -1.360858e-01 -0.0279154948 -0.14359321 -0.00929365
#>  [7,] -0.3320348 -1.563613e-01 -0.1051540830 -0.05983593 -0.11747026
#>  [8,] -0.2656916 -6.222279e-02  0.0354319025 -0.01361472 -0.01590426
#>  [9,] -0.3784858 -1.643468e-01 -0.0009895687 -0.07478936 -0.05382040
#> [10,] -0.2376183 -4.817807e-02 -0.1027185275 -0.04183129  0.01036119
#>                V14          V15         V16         V17          V18
#>  [1,] -0.015015633 -0.045034348  0.02865677 -0.02352236 -0.108797723
#>  [2,] -0.146492654 -0.011935167  0.06542265 -0.11496236 -0.022486284
#>  [3,] -0.071500968 -0.064317918 -0.15698746 -0.00926823  0.027293281
#>  [4,] -0.003442815 -0.086148337 -0.06483881 -0.03510471  0.114756813
#>  [5,] -0.047614051 -0.043030659  0.03153265  0.07587306  0.052712605
#>  [6,] -0.137554158  0.002277935  0.02493718  0.01125069  0.097041953
#>  [7,] -0.166807840 -0.101406141 -0.04606885 -0.01986665 -0.037229586
#>  [8,]  0.021716745  0.050906508 -0.03812578  0.02304125 -0.055252852
#>  [9,]  0.030755625 -0.028753487  0.06743536 -0.05103659  0.005743239
#> [10,]  0.023659122 -0.058628138 -0.07834286  0.06293924  0.083338476
#>                V19         V2          V20         V21         V22          V23
#>  [1,]  0.120299293 -0.3380287  0.129841144  0.02227784  0.12921067 -0.022778663
#>  [2,] -0.086186688 -0.1940755 -0.008344569 -0.10233661  0.09478123  0.067858539
#>  [3,] -0.005915894 -0.4634129  0.024276442  0.07035924  0.08999380  0.017608967
#>  [4,]  0.136664625 -0.1607245  0.048039240  0.06146853  0.01479141  0.083112842
#>  [5,]  0.157754977 -0.2492591  0.049956694  0.15907863  0.15036328  0.211987767
#>  [6,]  0.038678419 -0.2649539  0.004750636  0.02085433  0.12127852  0.142308128
#>  [7,] -0.102516671 -0.2329324 -0.160579135  0.02742582 -0.02281466  0.011850924
#>  [8,]  0.055005386 -0.3737799  0.185052738  0.18106159  0.10265770  0.068213958
#>  [9,]  0.049661757 -0.2388462  0.117683423  0.04363782  0.03915091  0.101705212
#> [10,] -0.105742989 -0.3544715  0.034879231  0.01088424  0.06697880  0.006066707
#>               V24          V25           V26          V27          V28
#>  [1,]  0.14241419  0.001990556  0.0005927526 -0.001259053 -0.040818879
#>  [2,]  0.11584318  0.026835427  0.0325347572  0.096793986 -0.073641386
#>  [3,]  0.13329615  0.096570094  0.0503443700  0.118735394  0.109635762
#>  [4,]  0.04932437  0.046239438  0.1463361172  0.065539439  0.055814610
#>  [5,]  0.05231965  0.064615738  0.1104707642  0.127402308  0.150672324
#>  [6,]  0.16592856  0.216285836  0.3182116475  0.248685036  0.188926697
#>  [7,] -0.08283011 -0.061650398 -0.0746186213 -0.051702229 -0.065791591
#>  [8,]  0.11385929  0.108000455  0.2906777914  0.250283062  0.237660464
#>  [9,]  0.15887192  0.115195496  0.0394031710  0.083121817 -0.007470595
#> [10,]  0.09481671  0.167782773  0.0055979336  0.133149808 -0.034408146
#>                V29         V3          V30          V31          V32
#>  [1,]  0.057690174 -0.2338232  0.105436263  0.038027502 -0.024219237
#>  [2,] -0.010336133 -0.3588286  0.105590462 -0.028969181  0.075357949
#>  [3,]  0.143065870 -0.4006635 -0.023457055 -0.021328621  0.002902352
#>  [4,]  0.111298433 -0.1886255  0.063437562 -0.015971458 -0.155859506
#>  [5,] -0.001108089 -0.2063063 -0.001586855 -0.006916424 -0.020895861
#>  [6,]  0.074678766 -0.3320137  0.050062525  0.125534031 -0.032414781
#>  [7,] -0.093272433 -0.3215328  0.026636268 -0.067643446 -0.068177774
#>  [8,]  0.089108025 -0.3370235  0.014596616  0.082047960 -0.060812079
#>  [9,] -0.037178428 -0.1977715 -0.001528288  0.037540130  0.083834145
#> [10,] -0.053892331 -0.2096856  0.013637945 -0.103267491 -0.059677474
#>               V33           V34          V35          V36         V37
#>  [1,]  0.08001975 -0.0303737173  0.064065873 -0.001808205  0.05844056
#>  [2,] -0.08792569 -0.0996565761 -0.147580454 -0.101359134 -0.06696753
#>  [3,] -0.07354448 -0.0396652263 -0.048211745 -0.027450278 -0.05300487
#>  [4,] -0.07355135 -0.0004000644 -0.052030294 -0.029798036 -0.12595700
#>  [5,] -0.03576615 -0.1078855202 -0.083069680 -0.162358062 -0.06773741
#>  [6,] -0.02856369 -0.1047031306  0.005165221  0.042410978  0.02072146
#>  [7,] -0.11583375 -0.1037757466 -0.194709489 -0.090189917 -0.04979432
#>  [8,]  0.01076344 -0.0034662502 -0.028186309 -0.102628572 -0.05464217
#>  [9,] -0.07773612  0.0359414085 -0.131659902 -0.088288088 -0.10591542
#> [10,] -0.15882633 -0.0668451451 -0.087851787 -0.114328467 -0.14003574
#>               V38         V39         V4          V40         V41         V42
#>  [1,]  0.08720239 -0.11216471 -0.2481541 -0.054534378 -0.06723311 -0.04117377
#>  [2,] -0.09762056 -0.01295001 -0.2132964  0.004657674  0.04986116  0.01584027
#>  [3,] -0.08793888 -0.05694308 -0.3381165 -0.068562505 -0.04970070 -0.11838834
#>  [4,] -0.15677784  0.02427274 -0.1487953 -0.095223826 -0.07990005 -0.16175169
#>  [5,] -0.09556384 -0.02542004 -0.2633274  0.070474950 -0.08030134 -0.12195426
#>  [6,]  0.02415860 -0.10564131 -0.2310409 -0.057702332 -0.06188669 -0.03663286
#>  [7,] -0.12198947 -0.09013654 -0.2917041 -0.110796567 -0.06430595 -0.05860761
#>  [8,] -0.01635998 -0.12163059 -0.2790481 -0.136093008 -0.10578577  0.01992956
#>  [9,] -0.10428334 -0.03377892 -0.1840814 -0.127888608 -0.13899267  0.01658999
#> [10,] -0.02488506 -0.13046229 -0.3140254 -0.075851970 -0.13809498 -0.13664593
#>               V43          V44         V45         V46        V47         V48
#>  [1,]  0.00549536 -0.130907730 -0.07651680 -0.17116307 -0.1618262 -0.06453579
#>  [2,] -0.14944082 -0.004478029 -0.15249571 -0.13625435 -0.1858827 -0.10904308
#>  [3,] -0.08659063 -0.256412105 -0.09904629 -0.23167331 -0.2477516 -0.35808410
#>  [4,] -0.16658153 -0.065174097 -0.03409336 -0.04795267 -0.2140436 -0.18475059
#>  [5,] -0.14292814 -0.119083284 -0.05745543 -0.16660060 -0.1338630 -0.11140187
#>  [6,] -0.03288108 -0.174003499 -0.18356738 -0.15332156 -0.1491236 -0.27869632
#>  [7,] -0.24514273 -0.106081653 -0.09229968 -0.10389456 -0.1674976 -0.18372698
#>  [8,] -0.17705442 -0.098960732 -0.16234561 -0.15831182 -0.1051465 -0.28088885
#>  [9,] -0.09302622 -0.029090831 -0.06871631 -0.11629844 -0.1113708 -0.27008800
#> [10,] -0.05131263 -0.052566403 -0.08012789 -0.07669439 -0.1316543 -0.10059835
#>              V49          V5        V50        V51        V52        V53
#>  [1,] -0.2714427 -0.20701945 -0.3203758 -0.3721102 -0.3230331 -0.3480467
#>  [2,] -0.2281260 -0.24212015 -0.3232919 -0.2323353 -0.3931185 -0.3961818
#>  [3,] -0.4074499 -0.28778276 -0.4084703 -0.5224363 -0.4273818 -0.5424943
#>  [4,] -0.2515583 -0.09031363 -0.2080941 -0.3057603 -0.2698302 -0.3302022
#>  [5,] -0.1179619 -0.26095871 -0.3564002 -0.2220549 -0.3342491 -0.2831112
#>  [6,] -0.3110804 -0.18072185 -0.4057659 -0.2967831 -0.3151418 -0.2856113
#>  [7,] -0.2628362 -0.26719342 -0.2835134 -0.2857015 -0.2489958 -0.2491223
#>  [8,] -0.3394340 -0.24308445 -0.3033283 -0.3556240 -0.3261647 -0.3742398
#>  [9,] -0.2842938 -0.24087902 -0.1860569 -0.2233588 -0.2186944 -0.3572754
#> [10,] -0.1569343 -0.30975182 -0.2521764 -0.3801812 -0.3177530 -0.2555948
#>              V54        V55        V56        V57        V58        V59
#>  [1,] -0.2795407 -0.3232851 -0.3115214 -0.3777767 -0.2817525 -0.3331921
#>  [2,] -0.3164795 -0.3913737 -0.2384392 -0.3722748 -0.3640283 -0.3709311
#>  [3,] -0.4862856 -0.4182716 -0.5155355 -0.4135663 -0.5097218 -0.4316405
#>  [4,] -0.3174556 -0.1945989 -0.2786779 -0.2062829 -0.1954089 -0.2552622
#>  [5,] -0.2831644 -0.2677381 -0.3228857 -0.3138122 -0.2883152 -0.2381808
#>  [6,] -0.4582041 -0.3099040 -0.4495106 -0.3272341 -0.3611291 -0.4341978
#>  [7,] -0.2595257 -0.4253270 -0.3923994 -0.3677419 -0.3579602 -0.3627618
#>  [8,] -0.3300113 -0.3981980 -0.3581616 -0.3575895 -0.3494204 -0.3646605
#>  [9,] -0.2326104 -0.3144329 -0.3381541 -0.2491171 -0.3778188 -0.3477544
#> [10,] -0.3035749 -0.2984702 -0.2164188 -0.3085025 -0.2543983 -0.3159449
#>                V6        V60          V7          V8          V9
#>  [1,] -0.14697118 -0.2942940 -0.22388965 -0.16204780  0.02633413
#>  [2,] -0.20673944 -0.4015068 -0.11263548 -0.19512547 -0.10145039
#>  [3,] -0.29507485 -0.5300965 -0.19709781 -0.14773527 -0.20183472
#>  [4,] -0.08204004 -0.3130086 -0.20577279 -0.10142522 -0.02469710
#>  [5,] -0.08668866 -0.3110685 -0.18558368 -0.19365397 -0.15609296
#>  [6,] -0.24431836 -0.2799418 -0.09529478 -0.23909778 -0.14846659
#>  [7,] -0.13758201 -0.3319202 -0.25282568 -0.12773969 -0.11562243
#>  [8,] -0.20877242 -0.3854396 -0.10576815 -0.05490949 -0.16126920
#>  [9,] -0.20637594 -0.3740925 -0.21114586 -0.10603699 -0.09743016
#> [10,] -0.27986303 -0.3498206 -0.20617829 -0.12318016 -0.13544455
#> 
#> $W[[2]]
#>              [,1]         [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.008485858 -0.009183389 -0.05986318  0.01033063  0.07216115  0.0182431
#> [2,] -0.041275071  0.002875535 -0.05217399 -0.04085278 -0.11870933 -0.0277906
#>            [,7]        [,8]        [,9]       [,10]
#> [1,] 0.05160872  0.03378522  0.11039564  0.11450331
#> [2,] 0.02494197 -0.20767504 -0.04938095 -0.08815358
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  4.042383e-04  0.0027537165  0.0030956576  0.0030212388  0.0032045831
#>  [2,]  4.065358e-04  0.0028224975  0.0031850669  0.0031115528  0.0032942691
#>  [3,]  3.613106e-04  0.0025343382  0.0028430081  0.0027828845  0.0029656874
#>  [4,]  3.037421e-04  0.0021031663  0.0023809430  0.0023572000  0.0024939446
#>  [5,]  6.178864e-05  0.0004812537  0.0006011718  0.0007209314  0.0007507084
#>  [6,]  3.876177e-04  0.0025028422  0.0027784763  0.0026526885  0.0028783731
#>  [7,]  1.779469e-05  0.0001370486  0.0001574780  0.0001734951  0.0001920092
#>  [8,]  9.064861e-06  0.0001504285  0.0002309388  0.0003745515  0.0003691986
#>  [9,]  5.247810e-05  0.0004255039  0.0005319508  0.0006524302  0.0006859231
#> [10,] -1.570785e-04 -0.0010138591 -0.0010990680 -0.0009138901 -0.0009410023
#>                 V14           V15           V16          V17          V18
#>  [1,]  0.0031483003  0.0030674717  0.0032988941 0.0033596550 0.0035562904
#>  [2,]  0.0031958556  0.0030838136  0.0032932449 0.0033521568 0.0035675696
#>  [3,]  0.0029459211  0.0029065331  0.0031595177 0.0032348543 0.0034214800
#>  [4,]  0.0024629499  0.0024208019  0.0026549621 0.0027678275 0.0029406294
#>  [5,]  0.0008085942  0.0008589761  0.0011067748 0.0013623049 0.0014761709
#>  [6,]  0.0029123950  0.0029092085  0.0031599881 0.0031746609 0.0032821432
#>  [7,]  0.0002247659  0.0002547711  0.0003177794 0.0003638804 0.0003835246
#>  [8,]  0.0004190816  0.0004613058  0.0006808118 0.0009337232 0.0010337394
#>  [9,]  0.0007735278  0.0008616853  0.0011329970 0.0013953182 0.0014970456
#> [10,] -0.0006984333 -0.0004498477 -0.0001529158 0.0001619083 0.0001549113
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0043631803  5.378930e-04  0.0051336969  0.0054788648  0.0052838492
#>  [2,]  0.0044016313  5.443472e-04  0.0051468346  0.0054665998  0.0052923692
#>  [3,]  0.0041184057  4.804203e-04  0.0047754599  0.0050904524  0.0049699467
#>  [4,]  0.0035406361  4.093134e-04  0.0041112483  0.0044017709  0.0042821810
#>  [5,]  0.0015834522  1.030967e-04  0.0016839090  0.0018829533  0.0019971082
#>  [6,]  0.0039913835  5.076406e-04  0.0047597292  0.0050608200  0.0047185248
#>  [7,]  0.0004019613  2.373205e-05  0.0004295964  0.0004779052  0.0005078471
#>  [8,]  0.0010342959  3.952452e-05  0.0010031739  0.0011577176  0.0013314606
#>  [9,]  0.0015596783  8.581133e-05  0.0016360758  0.0018350609  0.0019594866
#> [10,] -0.0002649355 -1.979562e-04 -0.0005763730 -0.0004757553 -0.0001979823
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0050075060 0.0050170757 0.0046719316 0.0048806609 0.0051912511
#>  [2,] 0.0049552692 0.0048989282 0.0044848420 0.0046538063 0.0050153506
#>  [3,] 0.0047484041 0.0047385281 0.0044366895 0.0046114576 0.0048869667
#>  [4,] 0.0041103881 0.0041511825 0.0038828811 0.0040426242 0.0042378681
#>  [5,] 0.0021914765 0.0023960194 0.0023678059 0.0024657219 0.0023845988
#>  [6,] 0.0044347689 0.0044646641 0.0041809795 0.0043907282 0.0046080683
#>  [7,] 0.0005559172 0.0005953159 0.0006147345 0.0006371383 0.0006151245
#>  [8,] 0.0015580735 0.0017415882 0.0017216455 0.0017809089 0.0016766379
#>  [9,] 0.0021731988 0.0023778134 0.0023976134 0.0024924599 0.0023822651
#> [10,] 0.0003477278 0.0006948085 0.0010400686 0.0010888697 0.0006847207
#>                 V28           V29            V3           V30          V31
#>  [1,]  5.726194e-03  0.0054851387  5.138694e-04  0.0049672171 0.0039622204
#>  [2,]  5.600610e-03  0.0053568545  5.086589e-04  0.0048165916 0.0037826830
#>  [3,]  5.315694e-03  0.0050638692  4.548274e-04  0.0045917051 0.0036839868
#>  [4,]  4.557050e-03  0.0043372587  4.093231e-04  0.0039262018 0.0031752935
#>  [5,]  2.150115e-03  0.0019822818  1.820600e-04  0.0017868849 0.0016495963
#>  [6,]  5.146178e-03  0.0049972433  4.863048e-04  0.0046319370 0.0036887705
#>  [7,]  5.623836e-04  0.0005125921  3.220371e-05  0.0004675414 0.0004220535
#>  [8,]  1.357941e-03  0.0012029243  1.253280e-04  0.0010785874 0.0010971317
#>  [9,]  2.102323e-03  0.0019134483  1.599194e-04  0.0017327499 0.0016216586
#> [10,] -7.531175e-05 -0.0002172455 -8.463116e-05 -0.0001373386 0.0002975681
#>                 V32          V33          V34          V35          V36
#>  [1,]  0.0036535027 0.0032906559 0.0026410457 0.0019419706 0.0016080895
#>  [2,]  0.0035216571 0.0031352419 0.0024014047 0.0016123512 0.0012554430
#>  [3,]  0.0033734768 0.0030617113 0.0025005208 0.0018933851 0.0016277402
#>  [4,]  0.0029071635 0.0026330429 0.0021565538 0.0016401035 0.0014066197
#>  [5,]  0.0013874327 0.0013107170 0.0013092192 0.0013231393 0.0013347719
#>  [6,]  0.0034285238 0.0031969971 0.0027241320 0.0021723490 0.0018571040
#>  [7,]  0.0003449601 0.0003330059 0.0003469636 0.0003617197 0.0003667367
#>  [8,]  0.0008792181 0.0008263964 0.0008657272 0.0009508408 0.0010263912
#>  [9,]  0.0013323570 0.0012748091 0.0013330259 0.0014062162 0.0014469102
#> [10,] -0.0000124803 0.0001608167 0.0007492574 0.0013839787 0.0016632904
#>                V37          V38           V39            V4          V40
#>  [1,] 0.0019419004 0.0026648043  0.0030436445  6.523944e-04 2.596545e-03
#>  [2,] 0.0016873539 0.0025219905  0.0029696487  6.481409e-04 2.537210e-03
#>  [3,] 0.0019288039 0.0024958679  0.0027856246  5.786790e-04 2.437506e-03
#>  [4,] 0.0016300422 0.0021502826  0.0024404573  5.111529e-04 2.109522e-03
#>  [5,] 0.0011662018 0.0010786744  0.0010991241  1.883357e-04 1.028667e-03
#>  [6,] 0.0020491104 0.0026486160  0.0029146702  6.212545e-04 2.482740e-03
#>  [7,] 0.0003286729 0.0002717204  0.0002437248  3.683476e-05 2.438083e-04
#>  [8,] 0.0008530488 0.0007032769  0.0007091445  1.094549e-04 7.018008e-04
#>  [9,] 0.0012497181 0.0010727103  0.0010325232  1.647733e-04 9.956642e-04
#> [10,] 0.0010677448 0.0001947979 -0.0002248923 -1.535772e-04 1.011875e-05
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0026425067  0.0029956773  0.0028010142  0.0026597148  0.0029049187
#>  [2,]  0.0026029327  0.0030018357  0.0028285598  0.0027017864  0.0029482171
#>  [3,]  0.0024433909  0.0027299046  0.0025506781  0.0024132665  0.0025800440
#>  [4,]  0.0021147435  0.0023420267  0.0021655544  0.0020492517  0.0022020944
#>  [5,]  0.0009416516  0.0008346859  0.0006795514  0.0006050698  0.0005241432
#>  [6,]  0.0025290757  0.0028153143  0.0025891368  0.0024292786  0.0027051020
#>  [7,]  0.0002186790  0.0001939771  0.0001652422  0.0001436802  0.0001155273
#>  [8,]  0.0005941970  0.0004337739  0.0003123872  0.0002672331  0.0001442157
#>  [9,]  0.0008876595  0.0007529638  0.0006034184  0.0005266441  0.0004149733
#> [10,] -0.0001984951 -0.0006484451 -0.0007638540 -0.0008101626 -0.0011331157
#>                 V46           V47           V48           V49            V5
#>  [1,]  2.373611e-03  1.694818e-03  1.333187e-03  7.830065e-04  0.0008395931
#>  [2,]  2.403411e-03  1.715713e-03  1.352601e-03  8.051570e-04  0.0008464346
#>  [3,]  2.087662e-03  1.501297e-03  1.187946e-03  7.020510e-04  0.0007562678
#>  [4,]  1.787238e-03  1.293256e-03  1.017306e-03  5.948610e-04  0.0006539591
#>  [5,]  4.046293e-04  3.591067e-04  2.703054e-04  1.355437e-04  0.0002221090
#>  [6,]  2.224215e-03  1.557741e-03  1.230252e-03  7.149188e-04  0.0007731030
#>  [7,]  9.040479e-05  7.805976e-05  5.910786e-05  3.056401e-05  0.0000479537
#>  [8,]  8.935559e-05  1.505624e-04  1.099757e-04  4.326920e-05  0.0001195662
#>  [9,]  3.129357e-04  2.959366e-04  2.206233e-04  1.059554e-04  0.0001966111
#> [10,] -9.575588e-04 -5.914583e-04 -4.877191e-04 -3.291859e-04 -0.0002221120
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.379453e-04  2.270203e-04  1.813931e-04  1.198445e-04  1.259411e-04
#>  [2,]  2.356990e-04  2.305751e-04  1.835203e-04  1.186420e-04  1.240773e-04
#>  [3,]  2.148681e-04  2.031276e-04  1.622340e-04  1.087007e-04  1.136060e-04
#>  [4,]  1.845022e-04  1.705711e-04  1.367590e-04  9.186153e-05  9.831327e-05
#>  [5,]  6.107771e-05  3.693595e-05  3.033196e-05  2.749124e-05  3.592973e-05
#>  [6,]  2.271893e-04  2.092665e-04  1.708469e-04  1.143842e-04  1.190584e-04
#>  [7,]  1.542798e-05  1.003126e-05  8.356922e-06  7.794614e-06  8.656621e-06
#>  [8,]  2.891906e-05  9.228887e-06  5.977811e-06  1.112935e-05  1.932539e-05
#>  [9,]  5.606359e-05  3.068478e-05  2.565975e-05  2.556766e-05  3.378887e-05
#> [10,] -5.329480e-05 -8.968741e-05 -6.842953e-05 -2.957945e-05 -2.257328e-05
#>                 V55           V56          V57           V58           V59
#>  [1,]  1.042241e-04  8.553993e-05 6.648881e-05  9.332759e-05  9.381089e-05
#>  [2,]  1.032746e-04  8.511316e-05 6.271013e-05  9.460686e-05  9.436940e-05
#>  [3,]  9.433796e-05  7.811660e-05 5.988853e-05  8.433417e-05  8.469368e-05
#>  [4,]  8.050448e-05  6.686879e-05 5.400230e-05  7.079220e-05  7.235896e-05
#>  [5,]  2.535391e-05  2.285949e-05 2.991578e-05  1.786838e-05  2.251945e-05
#>  [6,]  1.017560e-04  8.213155e-05 6.511386e-05  8.727697e-05  8.897494e-05
#>  [7,]  6.624895e-06  5.849614e-06 6.482376e-06  4.508114e-06  5.183537e-06
#>  [8,]  1.044026e-05  1.119911e-05 2.139459e-05  6.567001e-06  1.126363e-05
#>  [9,]  2.322810e-05  2.121348e-05 2.889145e-05  1.492394e-05  1.948879e-05
#> [10,] -2.522577e-05 -1.807818e-05 5.704783e-06 -3.425110e-05 -2.784541e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.057197e-03  6.871346e-05  1.177707e-03  1.538285e-03  2.368571e-03
#>  [2,]  1.069072e-03  6.836248e-05  1.176736e-03  1.562508e-03  2.425343e-03
#>  [3,]  9.868711e-04  6.243989e-05  1.104841e-03  1.430679e-03  2.174970e-03
#>  [4,]  8.214024e-04  5.372030e-05  9.235246e-04  1.178686e-03  1.801569e-03
#>  [5,]  2.752854e-04  1.975853e-05  3.452286e-04  2.975833e-04  3.848014e-04
#>  [6,]  9.482583e-04  6.497352e-05  1.072181e-03  1.411996e-03  2.170739e-03
#>  [7,]  7.515645e-05  4.567418e-06  9.509303e-05  9.300599e-05  1.143431e-04
#>  [8,]  1.514164e-04  1.219678e-05  2.087757e-04  1.094477e-04  9.197959e-05
#>  [9,]  2.625366e-04  1.776662e-05  3.363401e-04  2.825932e-04  3.393328e-04
#> [10,] -2.302299e-04 -1.443513e-05 -1.728681e-04 -4.662148e-04 -8.967521e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.06879999 -0.05137268 -0.04713532 -0.05857222 -0.07546225 -0.07849023
#> [2,]  0.06879999  0.05137268  0.04713532  0.05857222  0.07546225  0.07849023
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.01269653 -0.08828232 -0.06169654 -0.04498616
#> [2,]  0.01269653  0.08828232  0.06169654  0.04498616
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.5146941  0.5046084  0.5323782  0.3251777  0.4745895  0.5103012
#>  [7] -0.1154525  0.7191792  0.4090279  0.3936059
#> 
#> $B[[2]]
#> [1]  0.2854303 -0.2700912
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0079311133 0.0076962344 0.0074204129 0.0064236637 0.0032699453
#>  [6] 0.0073988433 0.0008423217 0.0022040388 0.0032303264 0.0004027741
#> 
#> $vB[[2]]
#> [1] -0.1269833  0.1269833
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#>  [2,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#>  [3,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>  [4,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [5,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#>  [6,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#>  [7,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [8,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#>  [9,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [10,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [11,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [12,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [13,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [14,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [15,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [16,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [17,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [18,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [19,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [20,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [21,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [22,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [23,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [24,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [25,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [26,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [27,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [28,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [29,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [30,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [31,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [32,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [33,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [34,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [35,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [36,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [37,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [38,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [39,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [40,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#>  [2,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#>  [3,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>  [4,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [5,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#>  [6,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#>  [7,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [8,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#>  [9,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [10,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [11,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [12,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [13,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [14,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [15,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [16,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [17,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [18,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [19,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [20,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [21,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [22,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [23,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [24,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [25,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [26,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [27,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [28,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [29,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [30,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [31,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [32,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [33,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [34,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [35,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [36,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [37,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [38,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [39,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [40,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#>  [2,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#>  [3,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>  [4,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [5,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#>  [6,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#>  [7,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [8,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#>  [9,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [10,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [11,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [12,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [13,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [14,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [15,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [16,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [17,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [18,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [19,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [20,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [21,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [22,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [23,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [24,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [25,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [26,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [27,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [28,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [29,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [30,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [31,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [32,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [33,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [34,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [35,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [36,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [37,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [38,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [39,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [40,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#>  [2,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#>  [3,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>  [4,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [5,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#>  [6,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#>  [7,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [8,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#>  [9,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [10,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [11,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [12,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [13,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [14,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [15,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [16,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [17,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [18,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [19,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [20,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [21,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [22,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [23,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [24,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [25,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [26,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [27,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [28,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [29,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [30,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [31,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [32,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [33,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [34,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [35,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [36,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [37,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [38,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [39,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [40,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#>  [2,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#>  [3,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>  [4,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [5,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#>  [6,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#>  [7,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [8,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#>  [9,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [10,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [11,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [12,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [13,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [14,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [15,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [16,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [17,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [18,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [19,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [20,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [21,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [22,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [23,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [24,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [25,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [26,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [27,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [28,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [29,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [30,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [31,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [32,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [33,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [34,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [35,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [36,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [37,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [38,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [39,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [40,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#>  [2,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#>  [3,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#>  [4,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [5,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#>  [6,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#>  [7,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [8,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#>  [9,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [10,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [11,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [12,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [13,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [14,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [15,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [16,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [17,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [18,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [19,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [20,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [21,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [22,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [23,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [24,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [25,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [26,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [27,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [28,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [29,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [30,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [31,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [32,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [33,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [34,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [35,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [36,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [37,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [38,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [39,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [40,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6038581 0.4511267 0.4629698 0.4689524 0.5377980 0.7157880 0.1313094
#>  [2,] 0.6142743 0.5273831 0.6095742 0.5888099 0.7006403 0.7574645 0.2451664
#>  [3,] 0.6114992 0.4954520 0.5454993 0.5636752 0.7107974 0.7448144 0.1518283
#>  [4,] 0.6308488 0.4699510 0.4751000 0.4995926 0.6617965 0.6510408 0.1800306
#>  [5,] 0.6685298 0.5958752 0.6689892 0.6345266 0.7398314 0.8399703 0.2238965
#>  [6,] 0.6666848 0.5486128 0.6203113 0.6430794 0.7642543 0.7922304 0.2051872
#>  [7,] 0.7005020 0.5679016 0.6554733 0.6134960 0.6931873 0.8036540 0.2155962
#>  [8,] 0.5015628 0.3756477 0.3305431 0.3859110 0.4672105 0.5636807 0.1162262
#>  [9,] 0.6802293 0.5866162 0.6604858 0.6101851 0.6861074 0.8348861 0.2070007
#> [10,] 0.6795042 0.5713605 0.6309583 0.6192156 0.7157897 0.8319213 0.1977179
#> [11,] 0.6306151 0.5260938 0.5718635 0.5699424 0.7193591 0.7932103 0.1680613
#> [12,] 0.6235318 0.5669677 0.6104170 0.5913509 0.7047506 0.8020032 0.2075783
#> [13,] 0.6771799 0.6072504 0.6655488 0.6381469 0.7373759 0.8462587 0.2324842
#> [14,] 0.5743961 0.4530977 0.4312343 0.4340708 0.5153011 0.6900753 0.1209832
#> [15,] 0.6894145 0.5568797 0.6426144 0.5958241 0.6812629 0.8256731 0.1937234
#> [16,] 0.5993402 0.3261587 0.3502173 0.4055083 0.4582932 0.5483122 0.1045824
#> [17,] 0.5865308 0.3673493 0.3824804 0.4371083 0.4983555 0.5422735 0.1384410
#> [18,] 0.5740286 0.4324853 0.4431581 0.5051087 0.6733763 0.6328018 0.1733165
#> [19,] 0.6086919 0.4744341 0.5032541 0.5849298 0.7221606 0.7514727 0.1370533
#> [20,] 0.5931385 0.5090003 0.4571611 0.4458529 0.5738658 0.6417005 0.1690928
#> [21,] 0.6752768 0.5780036 0.6718575 0.6660678 0.7696814 0.8401085 0.2083082
#> [22,] 0.6268243 0.5810954 0.6248909 0.6003205 0.7223095 0.8191253 0.2153967
#> [23,] 0.5841501 0.3281566 0.3517399 0.4352665 0.4754360 0.5379286 0.1042417
#> [24,] 0.6646746 0.5853194 0.6526390 0.6306541 0.7299628 0.8303027 0.2154319
#> [25,] 0.7135075 0.5886322 0.6983476 0.6484768 0.7410162 0.8616811 0.2103314
#> [26,] 0.5804405 0.4420173 0.4696154 0.5542319 0.7017668 0.6810997 0.1595818
#> [27,] 0.6791772 0.5343653 0.6027087 0.6420681 0.7573666 0.7760738 0.1986259
#> [28,] 0.5748756 0.4410668 0.4225211 0.4406585 0.5529863 0.6701681 0.1193061
#> [29,] 0.6454228 0.6034961 0.6570900 0.6210195 0.7248434 0.8355682 0.2307779
#> [30,] 0.6436921 0.5651129 0.6429187 0.6352286 0.7237194 0.8158176 0.2402637
#> [31,] 0.6904804 0.5616599 0.6347181 0.5839959 0.6557993 0.8152394 0.1786578
#> [32,] 0.6494174 0.5662871 0.6419038 0.6439720 0.7586792 0.8212639 0.2062879
#> [33,] 0.6409899 0.5190364 0.5805546 0.5930515 0.7124211 0.7668258 0.1826613
#> [34,] 0.6614839 0.5146386 0.5734869 0.5884216 0.7378141 0.7649322 0.1795488
#> [35,] 0.6825398 0.5850217 0.6544039 0.6615578 0.7926615 0.8381904 0.2165928
#> [36,] 0.6586303 0.5491958 0.5980652 0.5725386 0.6872634 0.7856221 0.1804010
#> [37,] 0.6919952 0.5641912 0.6395086 0.6347467 0.7407414 0.7745930 0.2475083
#> [38,] 0.6170866 0.5262736 0.5533830 0.5417971 0.6337119 0.7373151 0.1905733
#> [39,] 0.6922567 0.5572748 0.6325249 0.5831986 0.6603583 0.7995762 0.2165592
#> [40,] 0.6536072 0.5184654 0.5729990 0.5889162 0.7360156 0.7661495 0.1768521
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6985259 0.4644330 0.3510287
#>  [2,] 0.7985244 0.5810892 0.5445160
#>  [3,] 0.8056976 0.5812287 0.4670447
#>  [4,] 0.7115206 0.5656179 0.4471504
#>  [5,] 0.8537144 0.6493366 0.5584456
#>  [6,] 0.8370071 0.6422387 0.5282049
#>  [7,] 0.8128468 0.6072161 0.4814020
#>  [8,] 0.6101227 0.4339568 0.3187377
#>  [9,] 0.8269080 0.6220275 0.5065953
#> [10,] 0.8421002 0.6367176 0.5039155
#> [11,] 0.8293813 0.6248038 0.5183141
#> [12,] 0.8205445 0.6116121 0.5094118
#> [13,] 0.8528083 0.6638731 0.5626346
#> [14,] 0.6732834 0.4560152 0.3418393
#> [15,] 0.8204672 0.6013850 0.4780263
#> [16,] 0.6385370 0.4008609 0.2684800
#> [17,] 0.6313530 0.4299398 0.3392614
#> [18,] 0.7166450 0.5470848 0.4530385
#> [19,] 0.8019759 0.6049033 0.4937436
#> [20,] 0.6452149 0.5022061 0.3771448
#> [21,] 0.8696617 0.6681107 0.5690241
#> [22,] 0.8308235 0.6195405 0.5312056
#> [23,] 0.6255470 0.4038623 0.2882166
#> [24,] 0.8551520 0.6426882 0.5472352
#> [25,] 0.8655300 0.6530253 0.5357782
#> [26,] 0.7663875 0.5602131 0.4570414
#> [27,] 0.8240834 0.6432023 0.5147733
#> [28,] 0.7060187 0.5019307 0.3375921
#> [29,] 0.8414396 0.6394950 0.5378393
#> [30,] 0.8131614 0.6334853 0.5563556
#> [31,] 0.8163412 0.6119429 0.4592775
#> [32,] 0.8565969 0.6500706 0.5595365
#> [33,] 0.8250368 0.6299180 0.5225591
#> [34,] 0.7985361 0.6268183 0.5060419
#> [35,] 0.8598815 0.6913348 0.5821379
#> [36,] 0.8131599 0.6032915 0.4706615
#> [37,] 0.8055690 0.6587958 0.5695910
#> [38,] 0.7786845 0.5575668 0.4547978
#> [39,] 0.8038681 0.5980710 0.4842559
#> [40,] 0.7999314 0.6141101 0.4951685
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5380155 0.4619845
#>  [2,] 0.5384785 0.4615215
#>  [3,] 0.5398647 0.4601353
#>  [4,] 0.5428107 0.4571893
#>  [5,] 0.5328563 0.4671437
#>  [6,] 0.5357219 0.4642781
#>  [7,] 0.5308432 0.4691568
#>  [8,] 0.5484166 0.4515834
#>  [9,] 0.5308968 0.4691032
#> [10,] 0.5326306 0.4673694
#> [11,] 0.5386895 0.4613105
#> [12,] 0.5354926 0.4645074
#> [13,] 0.5323725 0.4676275
#> [14,] 0.5398442 0.4601558
#> [15,] 0.5313062 0.4686938
#> [16,] 0.5455535 0.4544465
#> [17,] 0.5460793 0.4539207
#> [18,] 0.5467659 0.4532341
#> [19,] 0.5420840 0.4579160
#> [20,] 0.5397951 0.4602049
#> [21,] 0.5336126 0.4663874
#> [22,] 0.5350420 0.4649580
#> [23,] 0.5461909 0.4538091
#> [24,] 0.5335518 0.4664482
#> [25,] 0.5299109 0.4700891
#> [26,] 0.5448350 0.4551650
#> [27,] 0.5360650 0.4639350
#> [28,] 0.5420327 0.4579673
#> [29,] 0.5328745 0.4671255
#> [30,] 0.5346878 0.4653122
#> [31,] 0.5310695 0.4689305
#> [32,] 0.5355141 0.4644859
#> [33,] 0.5387798 0.4612202
#> [34,] 0.5380118 0.4619882
#> [35,] 0.5344436 0.4655564
#> [36,] 0.5346959 0.4653041
#> [37,] 0.5353359 0.4646641
#> [38,] 0.5374600 0.4625400
#> [39,] 0.5320810 0.4679190
#> [40,] 0.5376628 0.4623372
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]        [,2]        [,3]         [,4]         [,5]      [,6]
#>  [1,] 0.42156688 -0.19611948 -0.14839247 -0.124350287  0.151480841 0.9236635
#>  [2,] 0.46531496  0.10964209  0.44552278  0.359047765  0.850348577 1.1388287
#>  [3,] 0.45361825 -0.01819245  0.18250216  0.256091419  0.899259766 1.0711438
#>  [4,] 0.53585983 -0.12034085 -0.09968253 -0.001629702  0.671310304 0.6236174
#>  [5,] 0.70154311  0.38830765  0.70361669  0.551685887  1.045092676 1.6580070
#>  [6,] 0.69322896  0.19506722  0.49086965  0.588754845  1.176146749 1.3384223
#>  [7,] 0.84968943  0.27329481  0.64318595  0.462031697  0.815062908 1.4092906
#>  [8,] 0.00625116 -0.50806321 -0.70572965 -0.464533119 -0.131346398 0.2561138
#>  [9,] 0.75482572  0.34999419  0.66545996  0.448090337  0.781983358 1.6206599
#> [10,] 0.75149414  0.28740394  0.53633014  0.486220161  0.923671847 1.5993059
#> [11,] 0.53485644  0.10447027  0.28945809  0.281616014  0.941284538 1.3443858
#> [12,] 0.50456625  0.26949021  0.44906555  0.369552889  0.870023379 1.3988616
#> [13,] 0.74084187  0.43576879  0.68812076  0.567330241  1.032374164 1.7055535
#> [14,] 0.29981013 -0.18816255 -0.27681696 -0.265261352  0.061223503 0.8004712
#> [15,] 0.79738346  0.22850814  0.58672969  0.388095134  0.759581691 1.5552664
#> [16,] 0.40271668 -0.72561034 -0.61808423 -0.382565242 -0.167215840 0.1938535
#> [17,] 0.34964235 -0.54360519 -0.47903333 -0.252906243 -0.006578193 0.1694985
#> [18,] 0.29830686 -0.27171821 -0.22835488  0.020435522  0.723495567 0.5442553
#> [19,] 0.44181714 -0.10235277  0.01301677  0.343043997  0.955204143 1.1064824
#> [20,] 0.37695498  0.03600501 -0.17177694 -0.217441059  0.297641201 0.5827523
#> [21,] 0.73214964  0.31458319  0.71659850  0.690453457  1.206512927 1.6590358
#> [22,] 0.51861673  0.32727161  0.51036023  0.406800874  0.955946752 1.5104323
#> [23,] 0.33983364 -0.71653421 -0.61140002 -0.260395442 -0.098334988 0.1520063
#> [24,] 0.68419606  0.34464898  0.63065958  0.535023741  0.994433793 1.5877738
#> [25,] 0.91248037  0.35831363  0.83944176  0.612350441  1.051256953 1.8293233
#> [26,] 0.32458205 -0.23297892 -0.12168853  0.217784371  0.855725296 0.7588304
#> [27,] 0.74999299  0.13767837  0.41676436  0.584351581  1.138295672 1.2429311
#> [28,] 0.30177195 -0.23683369 -0.31243235 -0.238489926  0.212744153 0.7089456
#> [29,] 0.59897943  0.42005355  0.65035289  0.493877649  0.968615244 1.6256158
#> [30,] 0.59142547  0.26193910  0.58805496  0.554714242  0.962986643 1.4882645
#> [31,] 0.80236616  0.24790155  0.55251177  0.339198734  0.644629751 1.4844212
#> [32,] 0.61647937  0.26671833  0.58363687  0.592645989  1.145451593 1.5249343
#> [33,] 0.57966327  0.07618238  0.32505038  0.376594493  0.907171947 1.1904739
#> [34,] 0.66991391  0.05857127  0.29609197  0.357444106  1.034638348 1.1799135
#> [35,] 0.76546818  0.34342270  0.63845382  0.670244148  1.341043281 1.6448250
#> [36,] 0.65719622  0.19742193  0.39740986  0.292216096  0.787356308 1.2987355
#> [37,] 0.80946381  0.25818959  0.57323181  0.552634908  1.049825595 1.2344300
#> [38,] 0.47720036  0.10519123  0.21434898  0.167579388  0.548174077 1.0320601
#> [39,] 0.81069062  0.23010907  0.54306410  0.335917728  0.664891255 1.3836479
#> [40,] 0.63493334  0.07389521  0.29409774  0.359486870  1.025361410 1.1866956
#>            [,7]      [,8]         [,9]       [,10]
#>  [1,] -1.889431 0.8402880 -0.142508507 -0.61452031
#>  [2,] -1.124560 1.3770972  0.327246300  0.17853660
#>  [3,] -1.720333 1.4222925  0.327819448 -0.13201262
#>  [4,] -1.516140 0.9027808  0.263994047 -0.21219087
#>  [5,] -1.243102 1.7640359  0.616124490  0.23485597
#>  [6,] -1.354184 1.6361261  0.585093985  0.11293937
#>  [7,] -1.291517 1.4686154  0.435625089 -0.07442630
#>  [8,] -2.028663 0.4478281 -0.265725509 -0.75957894
#>  [9,] -1.343100 1.5638701  0.498162789  0.02638270
#> [10,] -1.400619 1.6739387  0.561145531  0.01566236
#> [11,] -1.599430 1.5812485  0.509988704  0.07328930
#> [12,] -1.339585 1.5200407  0.454093742  0.03765159
#> [13,] -1.194337 1.7567993  0.680602073  0.25186124
#> [14,] -1.983152 0.7230734 -0.176395286 -0.65510847
#> [15,] -1.425995 1.5195163  0.411239408 -0.08795163
#> [16,] -2.147315 0.5690199 -0.401879430 -1.00234811
#> [17,] -1.828299 0.5380258 -0.282096598 -0.66658718
#> [18,] -1.562303 0.9278800  0.188899011 -0.18840129
#> [19,] -1.839983 1.3986896  0.425938241 -0.02502681
#> [20,] -1.592070 0.5980711  0.008824389 -0.50168478
#> [21,] -1.335153 1.8979710  0.699652315  0.27787062
#> [22,] -1.292697 1.5914750  0.487598410  0.12498482
#> [23,] -2.150958 0.5131600 -0.389397791 -0.90406116
#> [24,] -1.292488 1.7755945  0.587050731  0.18950577
#> [25,] -1.322929 1.8620009  0.632363852  0.14335798
#> [26,] -1.661343 1.1880244  0.242026927 -0.17225923
#> [27,] -1.394905 1.5442615  0.589290521  0.05911024
#> [28,] -1.999018 0.8761257  0.007722677 -0.67404318
#> [29,] -1.203924 1.6689787  0.573173163  0.15164704
#> [30,] -1.151234 1.4706843  0.547198061  0.22638420
#> [31,] -1.525468 1.4917530  0.455486380 -0.16325149
#> [32,] -1.347448 1.7873081  0.619349376  0.23928123
#> [33,] -1.498420 1.5508525  0.531864873  0.09029767
#> [34,] -1.519408 1.3771701  0.518591262  0.02416864
#> [35,] -1.285634 1.8143065  0.806367168  0.33155583
#> [36,] -1.513633 1.4706749  0.419198829 -0.11748890
#> [37,] -1.111946 1.4214713  0.657932480  0.28018258
#> [38,] -1.446289 1.2580169  0.231292991 -0.18130379
#> [39,] -1.285832 1.4106478  0.397433994 -0.06299712
#> [40,] -1.537822 1.3858655  0.464621976 -0.01932672
#> 
#> $pre[[3]]
#>               [,1]       [,2]
#>  [1,] -0.018538842 -0.1708950
#>  [2,] -0.030366139 -0.1845852
#>  [3,] -0.028822678 -0.1886205
#>  [4,] -0.008251919 -0.1799151
#>  [5,] -0.051913225 -0.1835281
#>  [6,] -0.043582606 -0.1867140
#>  [7,] -0.053390498 -0.1769202
#>  [8,]  0.022012248 -0.1722629
#>  [9,] -0.053819961 -0.1775647
#> [10,] -0.050695749 -0.1814041
#> [11,] -0.033195712 -0.1882637
#> [12,] -0.040004936 -0.1822144
#> [13,] -0.052065209 -0.1817368
#> [14,] -0.009754364 -0.1694697
#> [15,] -0.052710016 -0.1780988
#> [16,]  0.008250034 -0.1744705
#> [17,]  0.010046367 -0.1747952
#> [18,]  0.001586767 -0.1860253
#> [19,] -0.021639939 -0.1903750
#> [20,] -0.007104020 -0.1666217
#> [21,] -0.053113800 -0.1877675
#> [22,] -0.042540619 -0.1829388
#> [23,]  0.010486386 -0.1748054
#> [24,] -0.049935576 -0.1843446
#> [25,] -0.062644864 -0.1824316
#> [26,] -0.010017880 -0.1898407
#> [27,] -0.041194768 -0.1857058
#> [28,] -0.007281893 -0.1758103
#> [29,] -0.049324433 -0.1810122
#> [30,] -0.041937166 -0.1809117
#> [31,] -0.052104291 -0.1765427
#> [32,] -0.045867850 -0.1881639
#> [33,] -0.032653403 -0.1880849
#> [34,] -0.032937232 -0.1852784
#> [35,] -0.049077335 -0.1870705
#> [36,] -0.042121651 -0.1811287
#> [37,] -0.039950533 -0.1815302
#> [38,] -0.029226663 -0.1793478
#> [39,] -0.047558815 -0.1760594
#> [40,] -0.034155015 -0.1850921
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4619845 -0.4619845
#>  [2,] -0.5384785  0.5384785
#>  [3,]  0.4601353 -0.4601353
#>  [4,]  0.4571893 -0.4571893
#>  [5,] -0.5328563  0.5328563
#>  [6,]  0.4642781 -0.4642781
#>  [7,] -0.5308432  0.5308432
#>  [8,]  0.4515834 -0.4515834
#>  [9,] -0.5308968  0.5308968
#> [10,] -0.5326306  0.5326306
#> [11,]  0.4613105 -0.4613105
#> [12,]  0.4645074 -0.4645074
#> [13,]  0.4676275 -0.4676275
#> [14,]  0.4601558 -0.4601558
#> [15,] -0.5313062  0.5313062
#> [16,]  0.4544465 -0.4544465
#> [17,]  0.4539207 -0.4539207
#> [18,]  0.4532341 -0.4532341
#> [19,]  0.4579160 -0.4579160
#> [20,] -0.5397951  0.5397951
#> [21,]  0.4663874 -0.4663874
#> [22,]  0.4649580 -0.4649580
#> [23,]  0.4538091 -0.4538091
#> [24,]  0.4664482 -0.4664482
#> [25,] -0.5299109  0.5299109
#> [26,] -0.5448350  0.5448350
#> [27,]  0.4639350 -0.4639350
#> [28,]  0.4579673 -0.4579673
#> [29,]  0.4671255 -0.4671255
#> [30,] -0.5346878  0.5346878
#> [31,] -0.5310695  0.5310695
#> [32,]  0.4644859 -0.4644859
#> [33,] -0.5387798  0.5387798
#> [34,]  0.4619882 -0.4619882
#> [35,] -0.5344436  0.5344436
#> [36,]  0.4653041 -0.4653041
#> [37,] -0.5353359  0.5353359
#> [38,]  0.4625400 -0.4625400
#> [39,] -0.5320810  0.5320810
#> [40,]  0.4623372 -0.4623372
#> 
#> $L
#> [1] 0.6896366 0.7123531 1.0946825 0.6889177 0.7762508 0.6731052
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536