Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12           V13
#>  [1,] -0.2771858 -0.24428005 -0.265495990 -0.241573533 -0.1988593397
#>  [2,] -0.3800135 -0.11130173 -0.082269405 -0.055506300 -0.1103034676
#>  [3,] -0.3742554 -0.10193605 -0.005233607 -0.002814887 -0.0887132680
#>  [4,] -0.2105012 -0.06544992 -0.212012303 -0.153316587 -0.1674716877
#>  [5,] -0.3547415 -0.12876337 -0.091813480 -0.038659524 -0.1850266054
#>  [6,] -0.2967104 -0.06019869 -0.163722826 -0.109120667 -0.1486335193
#>  [7,] -0.2637337 -0.07658087 -0.066853559 -0.193131747 -0.0178497974
#>  [8,] -0.4324328 -0.22647479 -0.063338684 -0.021319652 -0.0993066653
#>  [9,] -0.2395751 -0.25143987 -0.129470968 -0.164876839 -0.0678432950
#> [10,] -0.2349644 -0.06759667 -0.130279249 -0.092327606 -0.0006515547
#>               V14         V15          V16          V17         V18         V19
#>  [1,] -0.17408075 -0.15382371 -0.213963778 -0.164918386 -0.08595011 -0.18969961
#>  [2,] -0.03390284 -0.04095142 -0.072461694  0.031241598  0.07127116  0.03427086
#>  [3,] -0.03958028 -0.05888370  0.049409431 -0.002827996 -0.02204365  0.05687276
#>  [4,] -0.03788839 -0.16776311 -0.096902281 -0.097425254 -0.05364583 -0.06613393
#>  [5,] -0.20735404 -0.11085572 -0.082224364  0.060726558  0.08594079  0.10778699
#>  [6,] -0.08406213 -0.12863878 -0.025337514 -0.047947291 -0.07238003  0.08029046
#>  [7,] -0.11470539 -0.08344720 -0.077872860  0.090909166  0.03199233  0.01313617
#>  [8,] -0.10360923 -0.02005065 -0.090644195 -0.136108433 -0.04183056  0.13002802
#>  [9,] -0.03763287 -0.02351819 -0.099126203 -0.024115475 -0.05198163 -0.12542534
#> [10,] -0.05432560  0.02674874 -0.005235892 -0.031142530  0.03875271 -0.03159537
#>               V2         V20         V21         V22        V23          V24
#>  [1,] -0.2362738 -0.01689080 -0.07549464  0.06104468 0.02591083 -0.007741108
#>  [2,] -0.2693437  0.02125845  0.17183948  0.16761694 0.21802811  0.230399005
#>  [3,] -0.3787874  0.05606651  0.17063877  0.20276661 0.16038948  0.253583269
#>  [4,] -0.2844687 -0.01120690 -0.04461058 -0.00958062 0.08624333  0.166819342
#>  [5,] -0.4021992  0.11498210  0.14943206  0.16630458 0.09288132  0.168691932
#>  [6,] -0.2665348  0.02037000 -0.03315426  0.01265010 0.08678647  0.092713516
#>  [7,] -0.3457907  0.02165627  0.12824136  0.10230838 0.18056460  0.070003057
#>  [8,] -0.3280262  0.07759351  0.11792065  0.19825177 0.12552951  0.077789915
#>  [9,] -0.2897321 -0.10657284  0.02561899 -0.09880456 0.10448533  0.152949785
#> [10,] -0.1782870  0.02362731  0.03069516  0.11764892 0.01814022  0.004399797
#>               V25         V26         V27         V28         V29         V3
#>  [1,] -0.06966103  0.07848924  0.02814429 -0.06316940 -0.03775413 -0.2188039
#>  [2,]  0.15345873  0.08533439  0.13847948  0.21193524  0.14256285 -0.3163849
#>  [3,]  0.15062735  0.24836440  0.21569174  0.20822428  0.13890981 -0.3307317
#>  [4,]  0.08871059  0.10830617  0.14097662  0.27612553  0.10375542 -0.1908162
#>  [5,]  0.14725453  0.09736738  0.08184173  0.12546266  0.08953474 -0.2528287
#>  [6,]  0.15396790  0.19568552  0.25989194  0.21345339  0.14747875 -0.2974841
#>  [7,]  0.11490151  0.15052879  0.24235244  0.10154925  0.11592870 -0.2448832
#>  [8,]  0.25419523  0.22187227  0.27585503  0.28188714  0.20591922 -0.4052215
#>  [9,]  0.13213005  0.23234577  0.13106417  0.12372282  0.24711061 -0.2703321
#> [10,]  0.04799237 -0.10425769 -0.09613644 -0.06916141 -0.02259558 -0.2477582
#>               V30          V31          V32         V33         V34
#>  [1,] -0.12290451 -0.023062921 -0.115078830  0.03068327 -0.07422973
#>  [2,]  0.12778635  0.007083886  0.048551518 -0.07087765 -0.03516897
#>  [3,]  0.03688451 -0.025486355 -0.036184571  0.01825349 -0.06456871
#>  [4,]  0.05258044  0.141488223 -0.055409103  0.03464678  0.12060562
#>  [5,]  0.05559901  0.064110839  0.006841219 -0.11323352 -0.11540739
#>  [6,]  0.07531353 -0.013864683  0.028785051  0.04132929  0.05105954
#>  [7,]  0.13249309  0.088156165 -0.052245270  0.11590691  0.03770021
#>  [8,]  0.18982174  0.174535118  0.070010536 -0.04816244  0.09583145
#>  [9,]  0.12119193  0.074189356 -0.009261923 -0.07565579 -0.04937112
#> [10,] -0.06935991  0.047695067  0.037703196  0.02503531 -0.11668053
#>                V35          V36          V37          V38         V39
#>  [1,] -0.132228681 -0.158506999 -0.094801148 -0.148175085 -0.13833745
#>  [2,] -0.076480618 -0.076518892 -0.192993067 -0.135381429 -0.07423142
#>  [3,]  0.055604975 -0.104173045 -0.001806963 -0.060344824 -0.15027826
#>  [4,]  0.004602663  0.035278541  0.022060028 -0.045950627  0.07740276
#>  [5,]  0.014831823 -0.073658827 -0.059300620 -0.194327990 -0.15088385
#>  [6,]  0.041056125  0.045302753 -0.069325016 -0.022942900 -0.06187311
#>  [7,]  0.100798974  0.091359975 -0.027306092  0.011191857 -0.11447546
#>  [8,]  0.013781311  0.016183077 -0.031636734 -0.023751478 -0.07124489
#>  [9,] -0.072985791 -0.126747012 -0.066624102 -0.098674786 -0.14401967
#> [10,] -0.023499216 -0.005249458 -0.053272072 -0.001775225  0.01604024
#>               V4         V40         V41         V42          V43         V44
#>  [1,] -0.1921853 -0.13872093 -0.07529614 -0.18230627 -0.200555575 -0.17645021
#>  [2,] -0.3753767 -0.13214439 -0.02374683 -0.11627353 -0.049241788 -0.05608175
#>  [3,] -0.2954787 -0.08770750 -0.05862009 -0.02802511 -0.008890905 -0.04667387
#>  [4,] -0.2804490  0.04498000 -0.10413195  0.02086758 -0.094665806 -0.14725319
#>  [5,] -0.2347285 -0.09076792 -0.19746707 -0.14723415 -0.200556040 -0.05930008
#>  [6,] -0.1521780 -0.04632963 -0.14129658  0.02127594  0.004463945 -0.15195914
#>  [7,] -0.2879437 -0.05250384 -0.04516694 -0.05602578 -0.093466135 -0.05266636
#>  [8,] -0.3303898 -0.12198730 -0.06484265 -0.07149199 -0.101359942 -0.07270563
#>  [9,] -0.3036779 -0.13012649 -0.05735988 -0.03063754 -0.101518393 -0.22502837
#> [10,] -0.2272289 -0.03853532 -0.02879553 -0.03432247 -0.080888788 -0.15019853
#>               V45         V46         V47         V48        V49         V5
#>  [1,] -0.15186653 -0.13211007 -0.14003660 -0.23580405 -0.2579595 -0.2318370
#>  [2,] -0.10599331 -0.17971176 -0.17929516 -0.28157333 -0.3972763 -0.1991608
#>  [3,] -0.16656212 -0.15761171 -0.15778724 -0.14531068 -0.2634981 -0.1936102
#>  [4,] -0.06227766 -0.12927350 -0.09730425 -0.09835748 -0.3115104 -0.1759326
#>  [5,] -0.20490164 -0.22172072 -0.28572592 -0.17305593 -0.2386426 -0.2552568
#>  [6,] -0.07924714 -0.03613779 -0.25088502 -0.22548960 -0.2559694 -0.2609165
#>  [7,] -0.18321310 -0.18210601 -0.09107144 -0.30672187 -0.2219710 -0.1621020
#>  [8,] -0.05049232 -0.18045897 -0.29412287 -0.27709541 -0.3830598 -0.3559194
#>  [9,] -0.12225507 -0.10900815 -0.19109956 -0.23813894 -0.2215579 -0.3389703
#> [10,] -0.05064396 -0.19551567 -0.16583024 -0.11558043 -0.2756592 -0.1504049
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3556525 -0.2803282 -0.3665803 -0.3823183 -0.2722193 -0.3316616
#>  [2,] -0.3457682 -0.3702138 -0.4430351 -0.2713074 -0.3036090 -0.3746108
#>  [3,] -0.3461572 -0.3084488 -0.3515825 -0.3522622 -0.4152280 -0.3913130
#>  [4,] -0.2900669 -0.2469654 -0.1958298 -0.1988925 -0.3087654 -0.3452229
#>  [5,] -0.3992336 -0.3504771 -0.2931882 -0.3242076 -0.4648336 -0.3441617
#>  [6,] -0.3104706 -0.3936156 -0.3643793 -0.3780287 -0.3018104 -0.3901582
#>  [7,] -0.2681622 -0.3598011 -0.2170697 -0.3090414 -0.2782395 -0.3242306
#>  [8,] -0.4269419 -0.3444273 -0.4733939 -0.4709849 -0.3969083 -0.3355847
#>  [9,] -0.3506344 -0.4029803 -0.2361098 -0.3937787 -0.2608972 -0.3804233
#> [10,] -0.2356714 -0.2196541 -0.2133963 -0.3408711 -0.3137451 -0.3489955
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3077432 -0.3159888 -0.3966837 -0.2287171 -0.1740987 -0.2677595
#>  [2,] -0.3509779 -0.3560680 -0.4492669 -0.3709374 -0.2172937 -0.4017092
#>  [3,] -0.4249736 -0.4325018 -0.2778773 -0.3454914 -0.1950931 -0.3860477
#>  [4,] -0.3488253 -0.2223943 -0.2311893 -0.2880013 -0.2698436 -0.2008783
#>  [5,] -0.3348244 -0.3002203 -0.4165480 -0.3543576 -0.1862038 -0.4145218
#>  [6,] -0.3203876 -0.4147523 -0.3506748 -0.4039892 -0.1147205 -0.2501476
#>  [7,] -0.3921183 -0.3974050 -0.3740345 -0.3518744 -0.1118588 -0.2968889
#>  [8,] -0.3180510 -0.4350572 -0.3737029 -0.3400046 -0.3103804 -0.3825520
#>  [9,] -0.2771185 -0.2515604 -0.2317465 -0.3536976 -0.1639606 -0.3499948
#> [10,] -0.3668738 -0.2864206 -0.2625318 -0.3545783 -0.2290702 -0.2925972
#>                V7          V8          V9
#>  [1,] -0.29644913 -0.14670459 -0.25222146
#>  [2,] -0.18182919 -0.18070938 -0.12554806
#>  [3,] -0.25129850 -0.08958316 -0.06680597
#>  [4,] -0.14136803 -0.13607799 -0.12690317
#>  [5,] -0.13708341 -0.28942517 -0.25982756
#>  [6,] -0.20647529 -0.15970664 -0.08084353
#>  [7,] -0.18133659 -0.24676170 -0.06089571
#>  [8,] -0.18062223 -0.23614586 -0.23512787
#>  [9,] -0.16960358 -0.12577990 -0.20705470
#> [10,] -0.09776353 -0.23703441 -0.06984027
#> 
#> $W[[2]]
#>             [,1]         [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.09306441 -0.061203582  0.10465495 -0.00781153 -0.07002399 0.05155883
#> [2,]  0.12770133  0.005967329 -0.04745081 -0.04377678 -0.06254609 0.09802470
#>              [,7]       [,8]        [,9]       [,10]
#> [1,]  0.005410676 0.05211981 -0.01482810 0.004157344
#> [2,] -0.041667608 0.04743729  0.02466741 0.065331180
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.397515e-04  0.0012572816  0.0015979214  0.0017747405  0.0015321088
#>  [2,]  2.496459e-04  0.0016220859  0.0019739267  0.0020967229  0.0020720666
#>  [3,] -2.421099e-05 -0.0001771757 -0.0003549877 -0.0002872220 -0.0001195689
#>  [4,]  4.352003e-05  0.0002290805  0.0001463947  0.0002275945  0.0003981643
#>  [5,]  1.300225e-04  0.0008433470  0.0009454158  0.0010618167  0.0011412439
#>  [6,]  2.003425e-04  0.0012709415  0.0015306387  0.0016712403  0.0017032175
#>  [7,]  5.864008e-05  0.0003470335  0.0003161601  0.0004001956  0.0005250991
#>  [8,]  1.244386e-04  0.0007204536  0.0008382841  0.0009195951  0.0009733712
#>  [9,]  1.647544e-04  0.0011054978  0.0012906690  0.0014308800  0.0014811270
#> [10,]  1.417624e-04  0.0009955752  0.0011793397  0.0013082106  0.0013190060
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0010158388 0.0007156583 0.0005053214 0.0004776692 0.0005138477
#>  [2,] 0.0017797463 0.0015841047 0.0015900342 0.0016207111 0.0016814141
#>  [3,] 0.0002535984 0.0005866747 0.0008715742 0.0010153515 0.0011315531
#>  [4,] 0.0007401164 0.0010277972 0.0013051854 0.0014485637 0.0015726854
#>  [5,] 0.0012098859 0.0012871997 0.0014423607 0.0015383439 0.0016482575
#>  [6,] 0.0015898916 0.0015321012 0.0016463067 0.0017583828 0.0018717368
#>  [7,] 0.0007462751 0.0009434778 0.0011570700 0.0012738573 0.0013765921
#>  [8,] 0.0009748307 0.0010051296 0.0011327909 0.0012257899 0.0012913054
#>  [9,] 0.0014568314 0.0014600036 0.0015963803 0.0017085437 0.0018375875
#> [10,] 0.0012423281 0.0012010966 0.0012657526 0.0013336103 0.0014489682
#>                V19            V2         V20          V21          V22
#>  [1,] 0.0013689927  2.228402e-04 0.001627512 0.0019636979 0.0018480823
#>  [2,] 0.0027346559  3.489488e-04 0.003307159 0.0037841866 0.0037955441
#>  [3,] 0.0007292664 -5.583715e-05 0.000710249 0.0006807851 0.0007884674
#>  [4,] 0.0014764371  4.384208e-05 0.001611623 0.0017013055 0.0017889521
#>  [5,] 0.0020406035  1.711904e-04 0.002372560 0.0026369715 0.0027075617
#>  [6,] 0.0026178301  2.697812e-04 0.003078430 0.0034748817 0.0035132927
#>  [7,] 0.0014011574  6.686872e-05 0.001565903 0.0016902135 0.0017569170
#>  [8,] 0.0016565618  1.608980e-04 0.001919845 0.0021518452 0.0021822859
#>  [9,] 0.0024414791  2.221535e-04 0.002854756 0.0031906204 0.0032516260
#> [10,] 0.0020317080  1.958470e-04 0.002393768 0.0026938897 0.0027181632
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.001360032 0.001624043 0.001753458 0.001577243 0.001661850 0.001778040
#>  [2,] 0.003468616 0.003804310 0.003917255 0.003870028 0.003961673 0.004188503
#>  [3,] 0.001052549 0.001008733 0.001144016 0.001384258 0.001386847 0.001344991
#>  [4,] 0.001912790 0.001998182 0.002173365 0.002334077 0.002300386 0.002242902
#>  [5,] 0.002672135 0.002876770 0.003025002 0.003126466 0.003165420 0.003239756
#>  [6,] 0.003308268 0.003572602 0.003690896 0.003696348 0.003731616 0.003864526
#>  [7,] 0.001840238 0.001942544 0.002095319 0.002231314 0.002229145 0.002227726
#>  [8,] 0.002108242 0.002267187 0.002369547 0.002404329 0.002420861 0.002507822
#>  [9,] 0.003111605 0.003346253 0.003481784 0.003535583 0.003574724 0.003658292
#> [10,] 0.002556076 0.002778832 0.002892157 0.002909766 0.002941131 0.003018490
#>               V29            V3          V30          V31          V32
#>  [1,] 0.001462300  2.554712e-04 0.0009346936 0.0003900014 0.0001668762
#>  [2,] 0.003801703  4.078818e-04 0.0031447598 0.0023810268 0.0019549088
#>  [3,] 0.001454874 -6.276716e-05 0.0014794452 0.0015166028 0.0012042182
#>  [4,] 0.002196967  6.010686e-05 0.0020406792 0.0019070083 0.0015250774
#>  [5,] 0.003053812  2.061003e-04 0.0026628440 0.0022429201 0.0018148124
#>  [6,] 0.003572429  3.192242e-04 0.0030124373 0.0023772900 0.0019238684
#>  [7,] 0.002158643  8.523720e-05 0.0019677200 0.0017772267 0.0014167377
#>  [8,] 0.002336910  1.946988e-04 0.0020266479 0.0016800604 0.0013665772
#>  [9,] 0.003400107  2.646378e-04 0.0028935755 0.0023387928 0.0018814604
#> [10,] 0.002783863  2.317379e-04 0.0023102566 0.0018064803 0.0014255095
#>                 V33           V34           V35           V36           V37
#>  [1,] -0.0003936105 -0.0007144689 -0.0011985306 -0.0017378464 -0.0011815393
#>  [2,]  0.0014849726  0.0012717645  0.0008132951  0.0001653350  0.0003485920
#>  [3,]  0.0015203518  0.0018069606  0.0021455305  0.0024505963  0.0019570249
#>  [4,]  0.0016496330  0.0018634951  0.0020627840  0.0021011221  0.0017029428
#>  [5,]  0.0016992423  0.0017353572  0.0016502127  0.0014070342  0.0012310219
#>  [6,]  0.0016526774  0.0015952439  0.0013228812  0.0008560717  0.0008417135
#>  [7,]  0.0014882226  0.0016407905  0.0017548975  0.0017456001  0.0014381128
#>  [8,]  0.0012623810  0.0012887899  0.0012086554  0.0009845884  0.0008713610
#>  [9,]  0.0016626989  0.0016456024  0.0014539831  0.0010870611  0.0010043646
#> [10,]  0.0011969052  0.0011502554  0.0009308646  0.0005631693  0.0005678152
#>                 V38          V39            V4          V40          V41
#>  [1,] -0.0002424059 0.0005985814  3.729005e-04 0.0005191743 0.0006378081
#>  [2,]  0.0009965788 0.0016443277  5.492156e-04 0.0015059622 0.0015747975
#>  [3,]  0.0011467082 0.0006542116 -1.115918e-04 0.0008741975 0.0006830935
#>  [4,]  0.0011054439 0.0008510311  6.250166e-05 0.0011035127 0.0009925789
#>  [5,]  0.0011421837 0.0012703114  2.624924e-04 0.0013655291 0.0013072703
#>  [6,]  0.0011308169 0.0015406723  4.246160e-04 0.0015121173 0.0014936777
#>  [7,]  0.0010345764 0.0008944879  9.826100e-05 0.0010624595 0.0009662990
#>  [8,]  0.0008818532 0.0010313640  2.560626e-04 0.0010626825 0.0010261968
#>  [9,]  0.0011257799 0.0014278503  3.460749e-04 0.0014749228 0.0014382993
#> [10,]  0.0007935968 0.0011436952  3.114678e-04 0.0011701643 0.0011556399
#>                V42          V43           V44           V45           V46
#>  [1,] 0.0010778439 0.0009942057  0.0011268577  1.429111e-03  0.0012157578
#>  [2,] 0.0019259000 0.0017910048  0.0017903214  2.007083e-03  0.0016198394
#>  [3,] 0.0002601744 0.0001568782 -0.0001065637 -4.622308e-04 -0.0003533005
#>  [4,] 0.0006767154 0.0005240167  0.0003042618  7.885084e-05  0.0000910201
#>  [5,] 0.0012675445 0.0011119511  0.0009573295  8.934200e-04  0.0007331219
#>  [6,] 0.0016467308 0.0014964972  0.0014008027  1.472538e-03  0.0011922592
#>  [7,] 0.0007494101 0.0006173960  0.0004303124  2.549063e-04  0.0002258005
#>  [8,] 0.0010282359 0.0009200054  0.0008210882  8.083109e-04  0.0006637524
#>  [9,] 0.0015034101 0.0013435950  0.0012258149  1.227717e-03  0.0009933139
#> [10,] 0.0012641288 0.0011329989  0.0010541718  1.098799e-03  0.0008967525
#>                 V47           V48           V49            V5          V50
#>  [1,]  0.0009353743  6.246084e-04  3.842419e-04  0.0006298144 8.633827e-05
#>  [2,]  0.0011598853  7.884893e-04  4.900878e-04  0.0008319261 1.483155e-04
#>  [3,] -0.0001978877 -1.249446e-04 -8.677321e-05 -0.0001500904 8.491527e-06
#>  [4,]  0.0001166243  8.617863e-05  4.656752e-05  0.0001105660 4.930809e-05
#>  [5,]  0.0005583529  3.882257e-04  2.367653e-04  0.0004171083 9.207215e-05
#>  [6,]  0.0008724783  5.952371e-04  3.706067e-04  0.0006577540 1.252555e-04
#>  [7,]  0.0002023530  1.421421e-04  8.389705e-05  0.0001635293 5.234910e-05
#>  [8,]  0.0004927705  3.329085e-04  2.060053e-04  0.0003829412 7.705586e-05
#>  [9,]  0.0007440197  5.145828e-04  3.177924e-04  0.0005560252 1.125311e-04
#> [10,]  0.0006756115  4.643043e-04  2.867459e-04  0.0005056733 9.587571e-05
#>                 V51           V52          V53          V54          V55
#>  [1,]  1.046511e-04  6.455884e-05 6.766019e-06 2.575901e-05 3.431088e-06
#>  [2,]  1.411664e-04  9.804898e-05 4.763303e-05 6.845128e-05 3.796967e-05
#>  [3,] -1.679575e-05 -1.620254e-05 2.739922e-05 1.237279e-05 1.063563e-05
#>  [4,]  1.900385e-05  9.753627e-06 3.413447e-05 2.719958e-05 2.003511e-05
#>  [5,]  7.217658e-05  4.778604e-05 4.241571e-05 4.596055e-05 2.812265e-05
#>  [6,]  1.097775e-04  7.636589e-05 4.811488e-05 6.192747e-05 3.498189e-05
#>  [7,]  2.873166e-05  1.653536e-05 3.271094e-05 2.877838e-05 1.889425e-05
#>  [8,]  6.217725e-05  4.245153e-05 3.337655e-05 4.016429e-05 2.288898e-05
#>  [9,]  9.435641e-05  6.395066e-05 4.612045e-05 5.445049e-05 3.170360e-05
#> [10,]  8.442567e-05  5.706974e-05 3.596472e-05 4.480171e-05 2.492837e-05
#>                V56          V57          V58          V59           V6
#>  [1,] 1.443592e-05 8.252899e-06 1.669983e-05 1.186795e-05 0.0004599731
#>  [2,] 3.848412e-05 3.603166e-05 4.207126e-05 4.119735e-05 0.0007436628
#>  [3,] 1.087545e-05 2.151696e-05 7.946479e-06 1.262280e-05 0.0000444085
#>  [4,] 1.901330e-05 2.925313e-05 1.558753e-05 1.964435e-05 0.0002391934
#>  [5,] 2.863653e-05 3.378587e-05 2.761454e-05 2.982119e-05 0.0004655128
#>  [6,] 3.498490e-05 3.739193e-05 3.710202e-05 3.768608e-05 0.0006358037
#>  [7,] 1.885368e-05 2.697831e-05 1.680126e-05 2.012060e-05 0.0002614500
#>  [8,] 2.206964e-05 2.665867e-05 2.371291e-05 2.529340e-05 0.0003891815
#>  [9,] 3.272074e-05 3.619506e-05 3.264880e-05 3.399738e-05 0.0005682625
#> [10,] 2.672186e-05 2.844912e-05 2.669544e-05 2.707966e-05 0.0004915142
#>                V60           V7           V8            V9
#>  [1,] 1.242169e-05 0.0003595149 4.020724e-04  0.0008610197
#>  [2,] 3.656595e-05 0.0007076306 7.294591e-04  0.0012339296
#>  [3,] 1.166126e-05 0.0001131578 8.571992e-05 -0.0001332663
#>  [4,] 1.969414e-05 0.0002945480 2.379705e-04  0.0001593046
#>  [5,] 2.771036e-05 0.0004787073 4.654841e-04  0.0006290653
#>  [6,] 3.411745e-05 0.0006249177 6.219695e-04  0.0009543262
#>  [7,] 1.931315e-05 0.0003018758 2.674742e-04  0.0002556155
#>  [8,] 2.294434e-05 0.0004005355 3.822399e-04  0.0005437318
#>  [9,] 3.131237e-05 0.0005633261 5.587159e-04  0.0008197323
#> [10,] 2.499428e-05 0.0004733378 4.759188e-04  0.0007296783
#> 
#> $vW[[2]]
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.006167305 -0.03379864 -0.03604092 -0.01504096 -0.01997062 -0.02553876
#> [2,] -0.006167305  0.03379864  0.03604092  0.01504096  0.01997062  0.02553876
#>             [,7]       [,8]        [,9]       [,10]
#> [1,] -0.02467076 -0.0339827 -0.01649993 -0.00628358
#> [2,]  0.02467076  0.0339827  0.01649993  0.00628358
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.1623534 0.4880439 0.5317642 0.3236435 0.5342070 0.5645934 0.5222911
#>  [8] 0.7412804 0.4981428 0.2003380
#> 
#> $B[[2]]
#> [1]  0.04716233 -0.07597854
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.001027828 0.004586136 0.002542818 0.003457470 0.004131599 0.004605418
#>  [7] 0.003224945 0.003169797 0.004473691 0.003542428
#> 
#> $vB[[2]]
#> [1] -0.04575741  0.04575741
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [2,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#>  [3,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#>  [4,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#>  [5,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#>  [6,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#>  [7,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [8,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#>  [9,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [10,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [11,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [12,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [13,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [14,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [15,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [16,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [17,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [18,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [19,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [20,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [21,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [22,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [23,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [24,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [25,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [26,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#> [27,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [28,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [29,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [30,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [31,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [32,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [33,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [34,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [35,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [36,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [37,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [38,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [39,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [40,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [2,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#>  [3,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#>  [4,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#>  [5,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#>  [6,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#>  [7,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [8,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#>  [9,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [10,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [11,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [12,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [13,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [14,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [15,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [16,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [17,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [18,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [19,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [20,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [21,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [22,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [23,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [24,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [25,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [26,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#> [27,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [28,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [29,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [30,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [31,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [32,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [33,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [34,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [35,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [36,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [37,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [38,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [39,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [40,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [2,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#>  [3,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#>  [4,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#>  [5,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#>  [6,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#>  [7,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [8,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#>  [9,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [10,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [11,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [12,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [13,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [14,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [15,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [16,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [17,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [18,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [19,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [20,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [21,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [22,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [23,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [24,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [25,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [26,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#> [27,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [28,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [29,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [30,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [31,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [32,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [33,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [34,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [35,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [36,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [37,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [38,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [39,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [40,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [2,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#>  [3,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#>  [4,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#>  [5,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#>  [6,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#>  [7,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [8,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#>  [9,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [10,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [11,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [12,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [13,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [14,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [15,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [16,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [17,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [18,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [19,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [20,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [21,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [22,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [23,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [24,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [25,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [26,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#> [27,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [28,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [29,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [30,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [31,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [32,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [33,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [34,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [35,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [36,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [37,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [38,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [39,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [40,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [2,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#>  [3,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#>  [4,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#>  [5,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#>  [6,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#>  [7,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [8,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#>  [9,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [10,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [11,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [12,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [13,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [14,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [15,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [16,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [17,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [18,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [19,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [20,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [21,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [22,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [23,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [24,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [25,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [26,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#> [27,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [28,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [29,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [30,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [31,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [32,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [33,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [34,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [35,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [36,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [37,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [38,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [39,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [40,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [2,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#>  [3,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#>  [4,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#>  [5,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#>  [6,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#>  [7,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [8,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#>  [9,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [10,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [11,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [12,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [13,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [14,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [15,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [16,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [17,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [18,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [19,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [20,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [21,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [22,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [23,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [24,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [25,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [26,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#> [27,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [28,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [29,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [30,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [31,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [32,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [33,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [34,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [35,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [36,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [37,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [38,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [39,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [40,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1031920 0.4927007 0.5874918 0.5432479 0.4061428 0.5947095 0.6557837
#>  [2,] 0.2667933 0.7569522 0.8049920 0.6905325 0.6665526 0.7665485 0.7621340
#>  [3,] 0.1320151 0.6929918 0.7867140 0.6976254 0.6135198 0.7712166 0.8105820
#>  [4,] 0.1703087 0.7037544 0.7790667 0.7148913 0.6117899 0.7718231 0.8116409
#>  [5,] 0.1630460 0.6502198 0.7388154 0.6837527 0.5176852 0.7488200 0.7480956
#>  [6,] 0.2167755 0.7781170 0.8309864 0.6423893 0.7154657 0.7407968 0.7827699
#>  [7,] 0.2380646 0.8297594 0.8717516 0.6477435 0.7824233 0.7739764 0.8211928
#>  [8,] 0.2306913 0.8099162 0.8526748 0.6613704 0.7345094 0.7726134 0.7878254
#>  [9,] 0.2359810 0.7414085 0.8093178 0.6628431 0.6616990 0.7478457 0.7779425
#> [10,] 0.2719311 0.7549267 0.8040685 0.6982407 0.6661303 0.7705222 0.7667796
#> [11,] 0.1350731 0.5950153 0.6903169 0.6620290 0.4533308 0.7091462 0.7123628
#> [12,] 0.1409502 0.6546426 0.7488718 0.5332236 0.5610145 0.6366187 0.7050083
#> [13,] 0.1975554 0.6775060 0.7538019 0.6386985 0.5514493 0.7020274 0.7399318
#> [14,] 0.2363407 0.7929478 0.8499815 0.7004328 0.7086903 0.7841860 0.8091673
#> [15,] 0.1230477 0.6666712 0.7339798 0.4551893 0.5673982 0.5929558 0.6637188
#> [16,] 0.2945692 0.7443281 0.7901996 0.6537697 0.6800830 0.7305652 0.7479422
#> [17,] 0.2770175 0.6279405 0.6707281 0.6584169 0.5649756 0.6856348 0.7039441
#> [18,] 0.2571284 0.7680096 0.8161161 0.7546351 0.6776814 0.8161515 0.8274245
#> [19,] 0.3053763 0.8017302 0.8446534 0.6960954 0.7362276 0.7883391 0.8001299
#> [20,] 0.2448735 0.7471506 0.7934327 0.6724135 0.6573037 0.7551077 0.7574265
#> [21,] 0.1533358 0.7722362 0.8287597 0.5826583 0.6812576 0.7182568 0.7700488
#> [22,] 0.2557952 0.8078504 0.8518110 0.6763743 0.7245046 0.7803441 0.7887920
#> [23,] 0.2043862 0.7080173 0.7566932 0.6011821 0.6688306 0.6984260 0.7353889
#> [24,] 0.2278901 0.7295120 0.7798919 0.6426395 0.6446589 0.7272409 0.7515138
#> [25,] 0.1591881 0.7012274 0.7679294 0.5246146 0.6433774 0.6449948 0.7347001
#> [26,] 0.1380549 0.6711474 0.7511190 0.5932047 0.5003354 0.6837661 0.6875423
#> [27,] 0.2585484 0.7558001 0.8181190 0.7310799 0.6867564 0.7955342 0.8197858
#> [28,] 0.2073582 0.8104244 0.8539927 0.6283076 0.7333581 0.7543870 0.7885847
#> [29,] 0.1607042 0.7353690 0.7963449 0.6068350 0.6146723 0.7079830 0.7255381
#> [30,] 0.2914379 0.8132028 0.8609237 0.7584100 0.7499519 0.8265928 0.8497745
#> [31,] 0.2029053 0.7829439 0.8367087 0.6348700 0.7065084 0.7425219 0.7836645
#> [32,] 0.1399369 0.6190923 0.6820783 0.4786412 0.5170440 0.5967612 0.6575453
#> [33,] 0.3269047 0.7633792 0.8115307 0.6967625 0.6915582 0.7674686 0.7659832
#> [34,] 0.1371208 0.6741768 0.7293915 0.6055268 0.5409297 0.6684901 0.7332879
#> [35,] 0.1428554 0.7325457 0.7997516 0.6086053 0.5980561 0.7182958 0.7308650
#> [36,] 0.3503919 0.8358117 0.8578802 0.7226416 0.7792005 0.8000400 0.8175466
#> [37,] 0.3265476 0.7665033 0.8082700 0.7290793 0.6922523 0.8003962 0.7988964
#> [38,] 0.1352846 0.6727730 0.7399888 0.5730454 0.5376675 0.6800393 0.7050140
#> [39,] 0.2351749 0.7516086 0.8062050 0.6637477 0.6557654 0.7504085 0.7678285
#> [40,] 0.1447839 0.5176842 0.5767056 0.4810875 0.4480128 0.5519715 0.5704312
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6991329 0.3923823 0.3951228
#>  [2,] 0.8599616 0.6953571 0.4250328
#>  [3,] 0.8705978 0.5869436 0.4009512
#>  [4,] 0.8705009 0.6235002 0.4144449
#>  [5,] 0.8329762 0.6013073 0.3479075
#>  [6,] 0.8532743 0.6077647 0.4494306
#>  [7,] 0.8866891 0.6580047 0.4821520
#>  [8,] 0.8788472 0.6651648 0.4176769
#>  [9,] 0.8524562 0.6206164 0.4428997
#> [10,] 0.8652723 0.6983326 0.4280795
#> [11,] 0.8021865 0.5340238 0.3433914
#> [12,] 0.7311623 0.5104491 0.3946766
#> [13,] 0.8025991 0.5474366 0.4284423
#> [14,] 0.8927598 0.6787966 0.4455970
#> [15,] 0.7179904 0.4518847 0.4053187
#> [16,] 0.8224518 0.6356740 0.4476060
#> [17,] 0.7696162 0.5761838 0.4571079
#> [18,] 0.9038628 0.6969417 0.4346343
#> [19,] 0.8864671 0.7138111 0.4712567
#> [20,] 0.8538676 0.6733270 0.4170427
#> [21,] 0.8369100 0.5835383 0.4099418
#> [22,] 0.8864934 0.6907297 0.4261248
#> [23,] 0.8142753 0.5699557 0.4599298
#> [24,] 0.8444182 0.6329926 0.4167149
#> [25,] 0.7838744 0.4763991 0.4520410
#> [26,] 0.7965716 0.5329222 0.3602017
#> [27,] 0.8893013 0.6724268 0.4459161
#> [28,] 0.8728563 0.6521795 0.4449160
#> [29,] 0.8343151 0.5498814 0.3911899
#> [30,] 0.9135587 0.7266208 0.4630495
#> [31,] 0.8563098 0.5995036 0.4497715
#> [32,] 0.7080063 0.4849425 0.4404560
#> [33,] 0.8551593 0.6815923 0.4671928
#> [34,] 0.7988065 0.4828567 0.4148647
#> [35,] 0.8456303 0.5520087 0.3874627
#> [36,] 0.8978534 0.7451342 0.4898278
#> [37,] 0.8784825 0.7285519 0.4504521
#> [38,] 0.8107610 0.5478048 0.3670370
#> [39,] 0.8490730 0.6303858 0.4312152
#> [40,] 0.6059522 0.4182552 0.3554602
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4534720 0.5465280
#>  [2,] 0.4251197 0.5748803
#>  [3,] 0.4364485 0.5635515
#>  [4,] 0.4331159 0.5668841
#>  [5,] 0.4382949 0.5617051
#>  [6,] 0.4286647 0.5713353
#>  [7,] 0.4230970 0.5769030
#>  [8,] 0.4255340 0.5744660
#>  [9,] 0.4288618 0.5711382
#> [10,] 0.4246356 0.5753644
#> [11,] 0.4440696 0.5559304
#> [12,] 0.4438410 0.5561590
#> [13,] 0.4365563 0.5634437
#> [14,] 0.4248602 0.5751398
#> [15,] 0.4460152 0.5539848
#> [16,] 0.4257181 0.5742819
#> [17,] 0.4322600 0.5677400
#> [18,] 0.4233309 0.5766691
#> [19,] 0.4195550 0.5804450
#> [20,] 0.4273195 0.5726805
#> [21,] 0.4343162 0.5656838
#> [22,] 0.4235363 0.5764637
#> [23,] 0.4327664 0.5672336
#> [24,] 0.4301448 0.5698552
#> [25,] 0.4392811 0.5607189
#> [26,] 0.4423613 0.5576387
#> [27,] 0.4245119 0.5754881
#> [28,] 0.4269643 0.5730357
#> [29,] 0.4365849 0.5634151
#> [30,] 0.4187018 0.5812982
#> [31,] 0.4294707 0.5705293
#> [32,] 0.4451632 0.5548368
#> [33,] 0.4212320 0.5787680
#> [34,] 0.4416310 0.5583690
#> [35,] 0.4374603 0.5625397
#> [36,] 0.4146098 0.5853902
#> [37,] 0.4194333 0.5805667
#> [38,] 0.4412322 0.5587678
#> [39,] 0.4286099 0.5713901
#> [40,] 0.4524951 0.5475049
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]        [,2]      [,3]        [,4]         [,5]      [,6]
#>  [1,] -2.1622500 -0.02919929 0.3536064  0.17342478 -0.379934261 0.3834691
#>  [2,] -1.0109537  1.13604188 1.4177920  0.80260998  0.692633945 1.1889234
#>  [3,] -1.8832576  0.81414364 1.3052306  0.83601570  0.462131878 1.2151937
#>  [4,] -1.5834411  0.86524073 1.2602359  0.91926021  0.454842421 1.2186343
#>  [5,] -1.6357367  0.62000551 1.0398208  0.77107189  0.070770299 1.0923290
#>  [6,] -1.2845573  1.25472672 1.5926341  0.58574960  0.922079991 1.0501138
#>  [7,] -1.1633200  1.58392284 1.9165356  0.60913507  1.279844194 1.2309020
#>  [8,] -1.2044119  1.44946564 1.7557358  0.66940698  1.017623346 1.2231271
#>  [9,] -1.1748416  1.05330240 1.4455837  0.67598971  0.670874785 1.0871554
#> [10,] -0.9848468  1.12506317 1.4119192  0.83893423  0.690734425 1.2112620
#> [11,] -1.8568293  0.38473769 0.8016014  0.67234916 -0.187221949 0.8912410
#> [12,] -1.8074201  0.63950941 1.0926041  0.13309051  0.245280261 0.5607180
#> [13,] -1.4016438  0.74233367 1.1189934  0.56971977  0.206528051 0.8569709
#> [14,] -1.1728476  1.34278663 1.7344563  0.84935963  0.889031524 1.2902293
#> [15,] -1.9638808  0.69316767 1.0149091 -0.17972491  0.271243569 0.3761982
#> [16,] -0.8732945  1.06858721 1.3261292  0.63565090  0.754153148 0.9974920
#> [17,] -0.9593042  0.52339143 0.7114799  0.65624732  0.261380451 0.7797897
#> [18,] -1.0609477  1.19710633 1.4902522  1.12348761  0.743136929 1.4904881
#> [19,] -0.8218255  1.39714330 1.6932678  0.82877307  1.026452877 1.3149429
#> [20,] -1.1261433  1.08347273 1.3457422  0.71912148  0.651301519 1.1260419
#> [21,] -1.7086735  1.22098107 1.5768624  0.33369542  0.759557070 0.9358310
#> [22,] -1.0679391  1.43610291 1.7488764  0.73715892  0.966917175 1.2676725
#> [23,] -1.3591023  0.88577375 1.1346345  0.41039278  0.702900923 0.8398137
#> [24,] -1.2202632  0.99214789 1.2650363  0.58683869  0.595643376 0.9806688
#> [25,] -1.6642813  0.85314927 1.1966563  0.09853824  0.590053237 0.5971096
#> [26,] -1.8315405  0.71337933 1.1045893  0.37722946  0.001341746 0.7711340
#> [27,] -1.0535271  1.12978983 1.5036555  1.00010842  0.784998599 1.3586130
#> [28,] -1.3409237  1.45277045 1.7662659  0.52496306  1.011727438 1.1221486
#> [29,] -1.6529976  1.02203585 1.3636043  0.43402770  0.466995071 0.8856080
#> [30,] -0.8884108  1.47095682 1.8229832  1.14398202  1.098356023 1.5616699
#> [31,] -1.3682342  1.28290537 1.6339406  0.55316688  0.878486190 1.0591179
#> [32,] -1.8158138  0.48569734 0.7633393 -0.08548712  0.068202397 0.3919880
#> [33,] -0.7222184  1.17129582 1.4599872  0.83192829  0.807413981 1.1940721
#> [34,] -1.8394123  0.72713737 0.9915375  0.42854773  0.164085863 0.7013639
#> [35,] -1.7917734  1.00757721 1.3847428  0.44145351  0.397372030 0.9360237
#> [36,] -0.6173169  1.62738959 1.7977942  0.95760275  1.261013384 1.3865446
#> [37,] -0.7238417  1.18867080 1.4388079  0.98995649  0.810670300 1.3887723
#> [38,] -1.8550194  0.72075369 1.0459105  0.29428732  0.150955927 0.7539525
#> [39,] -1.1793179  1.10721018 1.4255370  0.68004052  0.644479777 1.1007919
#> [40,] -1.7761116  0.07076649 0.3092641 -0.07568610 -0.208703152 0.2086397
#>            [,7]      [,8]        [,9]       [,10]
#>  [1,] 0.6445606 0.8431722 -0.43730923 -0.42582872
#>  [2,] 1.1644148 1.8149708  0.82528526 -0.30214655
#>  [3,] 1.4537963 1.9062545  0.35134458 -0.40150319
#>  [4,] 1.4607076 1.9053944  0.50443140 -0.34562017
#>  [5,] 1.0884812 1.6068688  0.41091499 -0.62824985
#>  [6,] 1.2818817 1.7605158  0.43792608 -0.20297177
#>  [7,] 1.5244500 2.0573591  0.65441503 -0.07142241
#>  [8,] 1.3118669 1.9815589  0.68639626 -0.33231715
#>  [9,] 1.2537163 1.7539966  0.49216543 -0.22940184
#> [10,] 1.1902157 1.8597884  0.83937060 -0.28969121
#> [11,] 0.9068872 1.4000167  0.13630586 -0.64821704
#> [12,] 0.8712622 1.0005275  0.04180266 -0.42769624
#> [13,] 1.0456140 1.4026188  0.19031896 -0.28820947
#> [14,] 1.4446086 2.1192462  0.74824680 -0.21847688
#> [15,] 0.6799108 0.9345150 -0.19305841 -0.38335197
#> [16,] 1.0876675 1.5330481  0.55663682 -0.21034812
#> [17,] 0.8661505 1.2061453  0.30712702 -0.17199093
#> [18,] 1.5674830 2.2409016  0.83277678 -0.26296794
#> [19,] 1.3871065 2.0551511  0.91396639 -0.11510019
#> [20,] 1.1386220 1.7652634  0.72327134 -0.33492520
#> [21,] 1.2085870 1.6354142  0.33731568 -0.36420616
#> [22,] 1.3176590 2.0554126  0.80353302 -0.29767951
#> [23,] 1.0221382 1.4780333  0.28167058 -0.16062538
#> [24,] 1.1067024 1.6914765  0.54507680 -0.33627372
#> [25,] 1.0186013 1.2883888 -0.09447386 -0.19242742
#> [26,] 0.7886545 1.3650027  0.13187949 -0.57448899
#> [27,] 1.5148966 2.0836241  0.71918173 -0.21718524
#> [28,] 1.3164152 1.9264527  0.62863320 -0.22123381
#> [29,] 0.9721012 1.6165237  0.20019161 -0.44231337
#> [30,] 1.7328336 2.3578824  0.97754512 -0.14807183
#> [31,] 1.2871506 1.7849726  0.40339724 -0.20159411
#> [32,] 0.6523743 0.8857206 -0.06024818 -0.23931173
#> [33,] 1.1857673 1.7756530  0.76109892 -0.13141765
#> [34,] 1.0113687 1.3788515 -0.06860002 -0.34389077
#> [35,] 0.9990158 1.7007323  0.20878996 -0.45799006
#> [36,] 1.4998130 2.1735979  1.07282728 -0.04069422
#> [37,] 1.3794112 1.9781381  0.98728785 -0.19884430
#> [38,] 0.8712898 1.4549628  0.19180517 -0.54494938
#> [39,] 1.1960898 1.7273487  0.53387239 -0.27689486
#> [40,] 0.2836105 0.4303289 -0.32994020 -0.59512309
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.06063336 0.1260188
#>  [2,] -0.11604734 0.1857438
#>  [3,] -0.09207384 0.1635145
#>  [4,] -0.09836782 0.1707815
#>  [5,] -0.08545060 0.1626344
#>  [6,] -0.11372637 0.1735750
#>  [7,] -0.12474771 0.1853251
#>  [8,] -0.11915246 0.1809436
#>  [9,] -0.10915054 0.1773458
#> [10,] -0.11649234 0.1872798
#> [11,] -0.07360004 0.1510619
#> [12,] -0.08490941 0.1406786
#> [13,] -0.09211715 0.1630330
#> [14,] -0.11712924 0.1857236
#> [15,] -0.08307719 0.1337071
#> [16,] -0.11817570 0.1811671
#> [17,] -0.10297446 0.1696617
#> [18,] -0.11720178 0.1919128
#> [19,] -0.12769077 0.1969095
#> [20,] -0.11236853 0.1804274
#> [21,] -0.10315954 0.1611027
#> [22,] -0.12113708 0.1871359
#> [23,] -0.10568421 0.1648890
#> [24,] -0.10784603 0.1734142
#> [25,] -0.09528087 0.1487994
#> [26,] -0.07986278 0.1517217
#> [27,] -0.11621565 0.1880629
#> [28,] -0.11647221 0.1777755
#> [29,] -0.09544449 0.1595895
#> [30,] -0.12896555 0.1991395
#> [31,] -0.11126828 0.1727428
#> [32,] -0.08066373 0.1395693
#> [33,] -0.12362122 0.1940965
#> [34,] -0.08528086 0.1492646
#> [35,] -0.09162436 0.1598514
#> [36,] -0.13925199 0.2056887
#> [37,] -0.12570139 0.1993989
#> [38,] -0.08409606 0.1520665
#> [39,] -0.10983088 0.1776941
#> [40,] -0.07159311 0.1190016
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4534720  0.4534720
#>  [2,]  0.5748803 -0.5748803
#>  [3,] -0.4364485  0.4364485
#>  [4,] -0.4331159  0.4331159
#>  [5,]  0.5617051 -0.5617051
#>  [6,]  0.5713353 -0.5713353
#>  [7,] -0.4230970  0.4230970
#>  [8,]  0.5744660 -0.5744660
#>  [9,] -0.4288618  0.4288618
#> [10,]  0.5753644 -0.5753644
#> [11,]  0.5559304 -0.5559304
#> [12,] -0.4438410  0.4438410
#> [13,] -0.4365563  0.4365563
#> [14,]  0.5751398 -0.5751398
#> [15,]  0.5539848 -0.5539848
#> [16,] -0.4257181  0.4257181
#> [17,] -0.4322600  0.4322600
#> [18,] -0.4233309  0.4233309
#> [19,] -0.4195550  0.4195550
#> [20,]  0.5726805 -0.5726805
#> [21,]  0.5656838 -0.5656838
#> [22,]  0.5764637 -0.5764637
#> [23,] -0.4327664  0.4327664
#> [24,]  0.5698552 -0.5698552
#> [25,]  0.5607189 -0.5607189
#> [26,]  0.5576387 -0.5576387
#> [27,] -0.4245119  0.4245119
#> [28,]  0.5730357 -0.5730357
#> [29,]  0.5634151 -0.5634151
#> [30,] -0.4187018  0.4187018
#> [31,]  0.5705293 -0.5705293
#> [32,] -0.4451632  0.4451632
#> [33,] -0.4212320  0.4212320
#> [34,]  0.5583690 -0.5583690
#> [35,]  0.5625397 -0.5625397
#> [36,] -0.4146098  0.4146098
#> [37,] -0.4194333  0.4194333
#> [38,]  0.5587678 -0.5587678
#> [39,]  0.5713901 -0.5713901
#> [40,]  0.5475049 -0.5475049
#> 
#> $L
#> [1] 0.7030080 0.7391340 0.7932769 0.8030002 0.7953819 0.7118573
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4202899