Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13         V14
#>  [1,] -0.2993685 -0.09542988 -0.09720858 -0.097814467 -0.00411203 -0.05863232
#>  [2,] -0.2451561 -0.18332415 -0.14596194 -0.074937202 -0.23206019 -0.05034318
#>  [3,] -0.3968151 -0.11355738  0.02562858 -0.149854016  0.03796514 -0.03034373
#>  [4,] -0.3150644 -0.23088082 -0.10294509 -0.025144814 -0.15351666 -0.08718775
#>  [5,] -0.2578807 -0.13017818 -0.16893057 -0.105676665 -0.12197419 -0.07822760
#>  [6,] -0.3325380 -0.08352039 -0.12329191 -0.129668018 -0.02610207 -0.12643521
#>  [7,] -0.3202753 -0.06059472 -0.14696129 -0.008485265 -0.04457979  0.04949838
#>  [8,] -0.3416662 -0.20042576 -0.06717446 -0.008521578 -0.10768178 -0.11850830
#>  [9,] -0.2543991 -0.08817852 -0.13458295 -0.139815739 -0.06537818 -0.07954381
#> [10,] -0.1959888 -0.05248871 -0.09879972 -0.111455737 -0.03789827 -0.05663239
#>                V15         V16         V17          V18           V19
#>  [1,]  0.005693977 -0.03868177  0.03636410  0.063402275  1.625694e-01
#>  [2,] -0.050407726 -0.16381665 -0.04899944 -0.042726712 -5.823992e-02
#>  [3,] -0.039415599  0.09939137  0.11446598  0.056428483  5.891913e-05
#>  [4,] -0.055135767 -0.11907417 -0.16182481 -0.144211003  2.197096e-02
#>  [5,] -0.057023645 -0.05532123 -0.01055188  0.073124183 -4.155794e-02
#>  [6,] -0.154149877 -0.13862524 -0.13084998 -0.077540262 -1.861858e-02
#>  [7,] -0.003010453  0.07885258  0.03047656  0.089626614  1.037283e-01
#>  [8,] -0.123619635  0.03715412  0.08176924  0.088191593  5.977005e-02
#>  [9,]  0.033228721  0.04604801  0.09600221  0.117842566  7.961168e-02
#> [10,] -0.040252229 -0.04441226  0.08306832 -0.008138174  5.314484e-02
#>               V2          V20         V21          V22         V23        V24
#>  [1,] -0.3003244 -0.009967666  0.09756782 -0.008088427  0.09789087 0.14330458
#>  [2,] -0.2128676  0.024290527 -0.04115467 -0.033958106 -0.03855218 0.03814934
#>  [3,] -0.3087856 -0.005536459  0.05104056  0.215698059  0.13539009 0.20034309
#>  [4,] -0.3400379 -0.029754104 -0.05432104 -0.041799196  0.09549094 0.10635895
#>  [5,] -0.3317452  0.078943814  0.04228373  0.078707659  0.11184576 0.22744364
#>  [6,] -0.3062101 -0.048191138  0.06159988  0.143780923  0.15590493 0.15896802
#>  [7,] -0.2445271  0.166584382  0.14832285  0.133154904  0.10406427 0.23643765
#>  [8,] -0.2185374 -0.004815688  0.08823922  0.155029313  0.10969211 0.20617520
#>  [9,] -0.3465674  0.167461218  0.06652569  0.214521412  0.17094953 0.13454390
#> [10,] -0.2453382  0.082631174  0.15150343  0.108357300  0.22397954 0.09706214
#>               V25        V26         V27         V28          V29         V3
#>  [1,]  0.16064210 0.14307915  0.06835803 0.171791523  0.123498181 -0.2575163
#>  [2,] -0.08147436 0.04877782 -0.02205991 0.076384331  0.004945379 -0.2220429
#>  [3,]  0.19252571 0.34597178  0.24012350 0.230933560  0.164927333 -0.3205502
#>  [4,]  0.03728085 0.11239578  0.08140190 0.178604274  0.190351317 -0.3692752
#>  [5,]  0.12769100 0.14113037  0.25168798 0.178718557  0.101477205 -0.1660591
#>  [6,]  0.12723506 0.21135611  0.09537549 0.108513115  0.163837004 -0.1835968
#>  [7,]  0.21107855 0.11531100  0.23883581 0.004394487  0.102472607 -0.2767269
#>  [8,]  0.21680387 0.17155276  0.21225755 0.105688595 -0.004751432 -0.3760155
#>  [9,]  0.24584272 0.21320115  0.18067985 0.190252372  0.171894824 -0.2660090
#> [10,]  0.12600395 0.12648921  0.16828115 0.056291130 -0.026949472 -0.1778222
#>                V30          V31          V32         V33          V34
#>  [1,]  0.027002467  0.032810068  0.126645262 -0.05844308 -0.112304074
#>  [2,] -0.097323274 -0.076544569  0.033731749 -0.12149978 -0.126074652
#>  [3,]  0.105142356  0.057336007  0.012021410 -0.05130717 -0.041847734
#>  [4,]  0.006798282  0.012579977 -0.009545062  0.03033842 -0.075290566
#>  [5,]  0.170065313  0.158172876  0.198401450  0.12202357  0.006183219
#>  [6,]  0.088383826  0.222980845  0.040740956  0.01031730  0.081598668
#>  [7,]  0.071504457 -0.075790628 -0.082882370 -0.11554572  0.009666467
#>  [8,] -0.075032837 -0.060451415 -0.142788864 -0.14993722 -0.074653059
#>  [9,]  0.012113043 -0.023538379  0.043675916 -0.03775815  0.015199149
#> [10,]  0.051977160  0.005376476 -0.057449246 -0.01136609 -0.066676156
#>                V35         V36         V37         V38         V39         V4
#>  [1,] -0.018879024 -0.08918566  0.04707471  0.02734006  0.05475226 -0.2162262
#>  [2,] -0.188469596 -0.11233175 -0.19361913 -0.13160607 -0.14703152 -0.2673900
#>  [3,] -0.113756849  0.02980439 -0.08934264 -0.09201152 -0.09847551 -0.2045484
#>  [4,]  0.097807429 -0.01081260 -0.03500969 -0.11880483 -0.10896302 -0.3315512
#>  [5,]  0.046571071  0.10728369  0.07487149  0.06543878  0.05278925 -0.2227887
#>  [6,]  0.069166460  0.06555031  0.05813648 -0.01535042 -0.04089685 -0.2941452
#>  [7,] -0.110701894 -0.11860715  0.04418212  0.01478405 -0.03548330 -0.2444488
#>  [8,] -0.066166111 -0.17209707 -0.22972242 -0.13621070 -0.06459780 -0.1617987
#>  [9,]  0.002878881  0.05999879 -0.09385729 -0.10985005 -0.06588540 -0.2980764
#> [10,] -0.188426576 -0.15823958 -0.18362529 -0.18432473 -0.24493643 -0.3086347
#>                V40          V41          V42         V43         V44
#>  [1,] -0.121309977 -0.095569219 -0.067833651  0.01354235 -0.15111941
#>  [2,] -0.138317194 -0.114227392 -0.092538141 -0.27415900 -0.29124930
#>  [3,] -0.060667773 -0.046907336 -0.039347008  0.03379546 -0.11362836
#>  [4,] -0.069370740 -0.080186418 -0.009130728 -0.03159570 -0.07059430
#>  [5,]  0.003377688  0.036153267 -0.085909034  0.01052090 -0.07616511
#>  [6,] -0.052553963 -0.027297998  0.003246046 -0.12757847 -0.14415196
#>  [7,] -0.059371497 -0.141580701 -0.141633358 -0.16491887 -0.16458195
#>  [8,] -0.148817595 -0.143395754 -0.193133838 -0.20395999 -0.06139149
#>  [9,] -0.037220553  0.005623456 -0.011912359 -0.16407725 -0.08741917
#> [10,] -0.056666953 -0.167817976 -0.152721493 -0.11766912 -0.05767436
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.17708219 -0.15507797 -0.1537199 -0.2697695 -0.2935341 -0.2677213
#>  [2,] -0.12910222 -0.13072806 -0.1819683 -0.2816751 -0.2579021 -0.2556478
#>  [3,] -0.13123279 -0.05819988 -0.1161961 -0.1288228 -0.1780055 -0.2025177
#>  [4,] -0.25097204 -0.14044538 -0.2105643 -0.2493611 -0.2149610 -0.2789101
#>  [5,] -0.15452524 -0.17615008 -0.1892747 -0.2529308 -0.2128641 -0.1705159
#>  [6,] -0.11789234 -0.13362751 -0.2480546 -0.1987919 -0.1846329 -0.1780944
#>  [7,] -0.20170442 -0.30559764 -0.2503591 -0.1740596 -0.4069359 -0.2158275
#>  [8,] -0.14805811 -0.18567585 -0.1675970 -0.1916593 -0.3605885 -0.2022899
#>  [9,] -0.05383449 -0.17364204 -0.1518609 -0.2704433 -0.2835863 -0.1909523
#> [10,] -0.12058183 -0.08476818 -0.2401962 -0.2429651 -0.1780557 -0.2965430
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2900015 -0.2691268 -0.3075002 -0.2261982 -0.3666027 -0.3169841
#>  [2,] -0.3696828 -0.4298352 -0.4004852 -0.4084704 -0.3245309 -0.3767584
#>  [3,] -0.3826967 -0.2788526 -0.2412970 -0.3754666 -0.3621574 -0.4273874
#>  [4,] -0.3708106 -0.2522551 -0.2581488 -0.2644032 -0.4023896 -0.3425608
#>  [5,] -0.3029984 -0.3403866 -0.3078905 -0.2673223 -0.2318672 -0.2579161
#>  [6,] -0.3077134 -0.3427516 -0.3451491 -0.2069392 -0.2647577 -0.2549200
#>  [7,] -0.2481297 -0.3120006 -0.2683443 -0.4238580 -0.4149531 -0.2887547
#>  [8,] -0.3580827 -0.2343824 -0.3752606 -0.3391668 -0.3443502 -0.2602897
#>  [9,] -0.3367428 -0.4160199 -0.4080301 -0.4099765 -0.3730756 -0.3543430
#> [10,] -0.3307139 -0.3140173 -0.2104405 -0.2058586 -0.2111144 -0.2966324
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3818478 -0.2316358 -0.3603439 -0.3465818 -0.20858488 -0.2052884
#>  [2,] -0.4060700 -0.4224511 -0.3448903 -0.2717188 -0.15424349 -0.2570330
#>  [3,] -0.3974999 -0.3209511 -0.4057443 -0.3674718 -0.26497002 -0.4078620
#>  [4,] -0.3316547 -0.3333094 -0.2830895 -0.3062358 -0.19500242 -0.3443921
#>  [5,] -0.3204494 -0.2475566 -0.3514993 -0.3224855 -0.14108957 -0.2318208
#>  [6,] -0.3097423 -0.2530534 -0.1996896 -0.3225043 -0.08508915 -0.3440121
#>  [7,] -0.3714596 -0.4225552 -0.3744739 -0.4218996 -0.28991839 -0.2914669
#>  [8,] -0.2365430 -0.3543459 -0.4083072 -0.3265783 -0.24253417 -0.3343227
#>  [9,] -0.3937441 -0.3659228 -0.4165470 -0.2682485 -0.16884586 -0.3896280
#> [10,] -0.3443289 -0.2521617 -0.3490282 -0.2842073 -0.23782540 -0.3056068
#>                V7          V8          V9
#>  [1,] -0.19691537 -0.23605320 -0.09516326
#>  [2,] -0.27406755 -0.29568577 -0.10168222
#>  [3,] -0.25722732 -0.15205321 -0.10972625
#>  [4,] -0.12764574 -0.24593192 -0.20778638
#>  [5,] -0.07360035 -0.07470709 -0.18945989
#>  [6,] -0.15823607 -0.08914132 -0.18116670
#>  [7,] -0.22544814 -0.15251621 -0.21786159
#>  [8,] -0.17590387 -0.13244706 -0.04258217
#>  [9,] -0.14111781 -0.18631598 -0.16732651
#> [10,] -0.09579540 -0.19114596 -0.11194026
#> 
#> $W[[2]]
#>            [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#> [1,] 0.03627042 -0.07796037 0.04551022 -0.0950542 -0.08566572 -0.05024994
#> [2,] 0.01372432  0.12352551 0.01877125  0.1636642  0.13473197  0.07597889
#>            [,7]         [,8]        [,9]      [,10]
#> [1,] -0.1200927 -0.104073819 -0.07686125 -0.0857594
#> [2,]  0.1170822  0.005002753 -0.01692660  0.0242407
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 9.281617e-05 0.0006533528 0.0006932692 0.0009349972 0.0019857961
#>  [2,] 3.081427e-05 0.0002839841 0.0004449274 0.0004868263 0.0001170432
#>  [3,] 6.837059e-05 0.0004888952 0.0006153330 0.0009775914 0.0017688315
#>  [4,] 1.453073e-04 0.0010124586 0.0013249106 0.0014062450 0.0011445494
#>  [5,] 1.526102e-04 0.0009594863 0.0012181092 0.0013374881 0.0015221797
#>  [6,] 1.224655e-04 0.0008538841 0.0010272262 0.0011834633 0.0017167369
#>  [7,] 1.833811e-04 0.0012156091 0.0015375675 0.0016302389 0.0015890165
#>  [8,] 1.683293e-04 0.0011290056 0.0013354541 0.0014534398 0.0018453134
#>  [9,] 9.771851e-05 0.0006604572 0.0008103107 0.0010578745 0.0016739069
#> [10,] 1.704681e-04 0.0011411552 0.0013598379 0.0014825594 0.0018394229
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0031203068  0.0032286877  0.0044265392  0.0055567210  0.0056664361
#>  [2,] -0.0003571100 -0.0004478456 -0.0007458348 -0.0008520156 -0.0007790761
#>  [3,]  0.0025222318  0.0024536024  0.0034307134  0.0044171092  0.0045518150
#>  [4,]  0.0006947725  0.0005419079  0.0003738695  0.0004355676  0.0006429737
#>  [5,]  0.0015791151  0.0015126999  0.0017784481  0.0021745328  0.0023454844
#>  [6,]  0.0021952759  0.0022045834  0.0028514667  0.0035082310  0.0036431689
#>  [7,]  0.0013425105  0.0011622481  0.0012389879  0.0015295028  0.0017123375
#>  [8,]  0.0021480757  0.0021145831  0.0026088481  0.0032001907  0.0033796032
#>  [9,]  0.0022546291  0.0022086070  0.0029744033  0.0037858482  0.0038907436
#> [10,]  0.0021001803  0.0020543974  0.0025176156  0.0030972906  0.0032777548
#>                 V19           V2           V20           V21           V22
#>  [1,]  0.0049629036 2.806713e-04  0.0046367848  0.0049012999  0.0061185774
#>  [2,] -0.0004658268 7.894581e-06 -0.0004531177 -0.0004996994 -0.0006860005
#>  [3,]  0.0040797166 2.206552e-04  0.0037384051  0.0038591925  0.0049276158
#>  [4,]  0.0012514203 1.525935e-04  0.0015230700  0.0016450026  0.0015891873
#>  [5,]  0.0025659955 2.244719e-04  0.0026631785  0.0028347145  0.0031261221
#>  [6,]  0.0034590516 2.400663e-04  0.0033615767  0.0035602801  0.0041663345
#>  [7,]  0.0021919530 2.401753e-04  0.0024400715  0.0026409118  0.0027896370
#>  [8,]  0.0034697942 2.764621e-04  0.0035888494  0.0038497726  0.0043420782
#>  [9,]  0.0035807176 2.288010e-04  0.0033644162  0.0035193656  0.0043626739
#> [10,]  0.0033971823 2.743060e-04  0.0035273265  0.0037953220  0.0042675465
#>                 V23          V24          V25          V26          V27
#>  [1,]  0.0070908569  0.007782077  0.008225522  0.008358908  0.008158607
#>  [2,] -0.0009096942 -0.001134863 -0.001401950 -0.001581912 -0.001488907
#>  [3,]  0.0058067283  0.006316745  0.006633460  0.006722394  0.006678434
#>  [4,]  0.0015092198  0.001421219  0.001171033  0.001070544  0.001174255
#>  [5,]  0.0033251035  0.003476307  0.003432539  0.003376792  0.003363562
#>  [6,]  0.0046492488  0.005046670  0.005216284  0.005202711  0.005089360
#>  [7,]  0.0028466956  0.002909678  0.002809628  0.002781016  0.002866541
#>  [8,]  0.0046908322  0.004994630  0.005076176  0.005115573  0.005077846
#>  [9,]  0.0050373102  0.005468379  0.005713199  0.005760888  0.005719291
#> [10,]  0.0045913656  0.004871317  0.004940079  0.004972560  0.004944354
#>                V28          V29            V3          V30           V31
#>  [1,]  0.007126633  0.006389014  3.923464e-04  0.005596247  0.0065147279
#>  [2,] -0.001328819 -0.001268384 -1.915562e-05 -0.001264957 -0.0015392610
#>  [3,]  0.005850987  0.005255442  3.157462e-04  0.004450571  0.0053751952
#>  [4,]  0.001284280  0.001084082  1.493048e-04  0.000675680  0.0001141428
#>  [5,]  0.003130439  0.002756460  2.588486e-04  0.002183824  0.0020818941
#>  [6,]  0.004507095  0.003959692  3.034834e-04  0.003329110  0.0036569989
#>  [7,]  0.002771473  0.002381519  2.521992e-04  0.001815559  0.0014517464
#>  [8,]  0.004650389  0.004107644  3.237150e-04  0.003448387  0.0034551216
#>  [9,]  0.005093918  0.004582498  3.014754e-04  0.003866242  0.0044003093
#> [10,]  0.004538264  0.003995934  3.192240e-04  0.003331243  0.0033030651
#>                V32           V33           V34           V35           V36
#>  [1,]  0.005608456  5.132046e-03  0.0053760966  0.0058424002  0.0066254886
#>  [2,] -0.001339950 -1.357637e-03 -0.0016020526 -0.0018797015 -0.0021114512
#>  [3,]  0.004676471  4.198801e-03  0.0043376524  0.0045350130  0.0052300296
#>  [4,]  0.000109923 -5.165539e-06 -0.0003552711 -0.0007719888 -0.0010834022
#>  [5,]  0.001803469  1.588416e-03  0.0014319340  0.0013040407  0.0012895354
#>  [6,]  0.003131529  2.772988e-03  0.0027554593  0.0028584691  0.0031811334
#>  [7,]  0.001308613  1.099342e-03  0.0007527767  0.0004433785  0.0002711806
#>  [8,]  0.003031257  2.773818e-03  0.0026607718  0.0026526031  0.0028158445
#>  [9,]  0.003789358  3.402606e-03  0.0034292478  0.0035922414  0.0040713341
#> [10,]  0.002893044  2.631744e-03  0.0025006255  0.0024720339  0.0026034814
#>                 V37           V38           V39           V4           V40
#>  [1,]  0.0057674479  0.0036327456  0.0024840537 0.0002253240  0.0032455659
#>  [2,] -0.0015827421 -0.0008354738 -0.0003747969 0.0000583923 -0.0006010646
#>  [3,]  0.0046353466  0.0029863083  0.0021326522 0.0001914946  0.0026789839
#>  [4,] -0.0005396603  0.0002092895  0.0006445711 0.0002476556  0.0003209449
#>  [5,]  0.0014080313  0.0013020368  0.0012494050 0.0002849086  0.0012274316
#>  [6,]  0.0029631374  0.0021110267  0.0016549493 0.0002554385  0.0019520137
#>  [7,]  0.0006994612  0.0010363725  0.0012228300 0.0003233864  0.0010035302
#>  [8,]  0.0027710164  0.0022133503  0.0018685109 0.0003084549  0.0019544203
#>  [9,]  0.0036854632  0.0024926962  0.0018376739 0.0002156222  0.0022119418
#> [10,]  0.0026028907  0.0021213848  0.0018186168 0.0003107666  0.0018799677
#>                 V41           V42           V43          V44          V45
#>  [1,]  0.0028946123  0.0018870987  1.685585e-03 1.214626e-03 0.0001950489
#>  [2,] -0.0005483011 -0.0001143173 -1.183767e-05 8.337821e-05 0.0003398215
#>  [3,]  0.0024049148  0.0016478948  1.521139e-03 1.115760e-03 0.0002132463
#>  [4,]  0.0003592549  0.0008351724  9.128115e-04 9.393882e-04 0.0011527744
#>  [5,]  0.0011562665  0.0012154456  1.189589e-03 1.052227e-03 0.0009233242
#>  [6,]  0.0017845454  0.0014376849  1.326537e-03 1.102319e-03 0.0006502498
#>  [7,]  0.0009975407  0.0013088408  1.310733e-03 1.275390e-03 0.0013770018
#>  [8,]  0.0017937810  0.0016106061  1.527679e-03 1.320934e-03 0.0010210036
#>  [9,]  0.0020156892  0.0015147252  1.413410e-03 1.088359e-03 0.0004537469
#> [10,]  0.0017348070  0.0015939112  1.517612e-03 1.322571e-03 0.0010513114
#>                V46          V47          V48          V49           V5
#>  [1,] 0.0001760179 0.0003359250 0.0003255898 0.0001872492 4.453781e-04
#>  [2,] 0.0002636654 0.0001738552 0.0001134306 0.0000733277 5.489972e-05
#>  [3,] 0.0001668613 0.0002861569 0.0002722894 0.0001718463 4.017819e-04
#>  [4,] 0.0009023286 0.0006716552 0.0005113120 0.0002869605 3.128054e-04
#>  [5,] 0.0007434617 0.0006212139 0.0005015240 0.0002775672 3.844552e-04
#>  [6,] 0.0005287317 0.0005143599 0.0004346719 0.0002404899 4.035118e-04
#>  [7,] 0.0010870859 0.0008420647 0.0006691473 0.0003692190 4.182564e-04
#>  [8,] 0.0008082009 0.0006937147 0.0005735118 0.0003186406 4.545103e-04
#>  [9,] 0.0003701264 0.0004078154 0.0003621564 0.0002106156 3.851704e-04
#> [10,] 0.0008326403 0.0007082374 0.0005822652 0.0003232003 4.550347e-04
#>                 V50          V51           V52           V53           V54
#>  [1,]  1.745662e-04 8.209104e-05  1.139368e-04  1.278837e-04  1.150467e-04
#>  [2,] -3.481682e-05 5.771132e-06 -1.240443e-05 -3.221310e-05 -2.246440e-05
#>  [3,]  1.360654e-04 6.873935e-05  9.474146e-05  1.038424e-04  8.886994e-05
#>  [4,]  4.250158e-05 7.599242e-05  4.420461e-05  1.173054e-05  2.410702e-05
#>  [5,]  9.099399e-05 8.769567e-05  7.326407e-05  5.160252e-05  5.807178e-05
#>  [6,]  1.201092e-04 8.352056e-05  8.595635e-05  7.584559e-05  7.791099e-05
#>  [7,]  8.987328e-05 1.004914e-04  7.549076e-05  3.960493e-05  5.077745e-05
#>  [8,]  1.317335e-04 1.012186e-04  9.413260e-05  7.656517e-05  8.275216e-05
#>  [9,]  1.244009e-04 7.580293e-05  8.996013e-05  8.829848e-05  8.036285e-05
#> [10,]  1.287915e-04 1.015720e-04  9.288351e-05  7.423728e-05  8.057457e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  8.279645e-05  7.389059e-05  9.057632e-05  5.421703e-05  5.173557e-05
#>  [2,] -1.547450e-05 -1.391273e-05 -2.425913e-05 -5.448295e-06 -7.986137e-06
#>  [3,]  6.333039e-05  5.702428e-05  6.802391e-05  4.508261e-05  4.161623e-05
#>  [4,]  2.082908e-05  1.547777e-05  3.575727e-07  2.261003e-05  1.710496e-05
#>  [5,]  4.620463e-05  3.610023e-05  2.933836e-05  3.661192e-05  3.132878e-05
#>  [6,]  5.859820e-05  4.950526e-05  5.098006e-05  4.012264e-05  3.698649e-05
#>  [7,]  4.225299e-05  2.993787e-05  1.807213e-05  3.458721e-05  2.803060e-05
#>  [8,]  6.298059e-05  5.052803e-05  4.936791e-05  4.564739e-05  4.067206e-05
#>  [9,]  6.086697e-05  5.005800e-05  5.596617e-05  4.268878e-05  3.924098e-05
#> [10,]  6.166944e-05  4.940559e-05  4.707258e-05  4.513720e-05  3.994817e-05
#>                  V6           V60            V7            V8           V9
#>  [1,]  0.0012095116  3.229522e-05  0.0011423854  9.491635e-04 0.0006199472
#>  [2,] -0.0001535536 -5.870493e-06 -0.0001496267 -6.904172e-05 0.0001622443
#>  [3,]  0.0010169746  1.972102e-05  0.0008961816  7.173481e-04 0.0004496356
#>  [4,]  0.0001918572  1.465862e-05  0.0002665755  3.814653e-04 0.0008101932
#>  [5,]  0.0005204776  2.337469e-05  0.0005712960  5.855471e-04 0.0007968327
#>  [6,]  0.0007804182  2.547280e-05  0.0007803404  7.267252e-04 0.0007513312
#>  [7,]  0.0004219319  2.197514e-05  0.0005135856  5.724691e-04 0.0009906995
#>  [8,]  0.0007588068  3.060305e-05  0.0008180253  7.924208e-04 0.0009635811
#>  [9,]  0.0008516847  2.185496e-05  0.0008090407  6.993866e-04 0.0005882430
#> [10,]  0.0007365390  3.016412e-05  0.0007975130  7.795839e-04 0.0009704616
#> 
#> $vW[[2]]
#>            [,1]        [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,]  0.1229644  0.05189587  0.1348905  0.1016893  0.1358994  0.1288188
#> [2,] -0.1229644 -0.05189587 -0.1348905 -0.1016893 -0.1358994 -0.1288188
#>            [,7]        [,8]       [,9]       [,10]
#> [1,]  0.1177197  0.09824468  0.1334541  0.09501511
#> [2,] -0.1177197 -0.09824468 -0.1334541 -0.09501511
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5323737 0.2105645 0.6342105 0.4888092 0.4249851 0.5580904 0.4483847
#>  [8] 0.4294551 0.5367379 0.3546815
#> 
#> $B[[2]]
#> [1] -0.02558478  0.10717658
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.011562309 -0.002398941  0.009324431  0.001101080  0.004508899
#>  [6]  0.007014048  0.003452256  0.006867661  0.007942410  0.006639223
#> 
#> $vB[[2]]
#> [1]  0.1594823 -0.1594823
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#>  [2,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#>  [3,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#>  [4,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [5,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#>  [6,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#>  [7,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#>  [8,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [9,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [10,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [11,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [12,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [13,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [14,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [15,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [16,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#> [17,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [18,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [19,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [20,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [21,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [22,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [23,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [24,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [25,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [26,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [27,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [28,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [29,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [30,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [31,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [32,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [33,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [34,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [35,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [36,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [37,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [38,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [39,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [40,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#>  [2,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#>  [3,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#>  [4,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [5,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#>  [6,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#>  [7,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#>  [8,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [9,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [10,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [11,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [12,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [13,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [14,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [15,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [16,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#> [17,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [18,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [19,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [20,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [21,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [22,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [23,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [24,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [25,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [26,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [27,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [28,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [29,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [30,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [31,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [32,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [33,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [34,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [35,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [36,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [37,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [38,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [39,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [40,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#>  [2,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#>  [3,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#>  [4,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [5,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#>  [6,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#>  [7,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#>  [8,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [9,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [10,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [11,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [12,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [13,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [14,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [15,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [16,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#> [17,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [18,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [19,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [20,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [21,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [22,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [23,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [24,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [25,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [26,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [27,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [28,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [29,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [30,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [31,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [32,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [33,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [34,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [35,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [36,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [37,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [38,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [39,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [40,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#>  [2,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#>  [3,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#>  [4,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [5,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#>  [6,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#>  [7,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#>  [8,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [9,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [10,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [11,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [12,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [13,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [14,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [15,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [16,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#> [17,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [18,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [19,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [20,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [21,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [22,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [23,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [24,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [25,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [26,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [27,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [28,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [29,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [30,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [31,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [32,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [33,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [34,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [35,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [36,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [37,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [38,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [39,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [40,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#>  [2,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#>  [3,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#>  [4,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [5,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#>  [6,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#>  [7,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#>  [8,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [9,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [10,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [11,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [12,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [13,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [14,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [15,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [16,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#> [17,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [18,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [19,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [20,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [21,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [22,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [23,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [24,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [25,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [26,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [27,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [28,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [29,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [30,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [31,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [32,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [33,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [34,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [35,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [36,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [37,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [38,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [39,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [40,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#>  [2,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#>  [3,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#>  [4,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [5,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#>  [6,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#>  [7,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#>  [8,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [9,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [10,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [11,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [12,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [13,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [14,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [15,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [16,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#> [17,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [18,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [19,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [20,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [21,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [22,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [23,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [24,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [25,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [26,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [27,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [28,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [29,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [30,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [31,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [32,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [33,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [34,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [35,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [36,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [37,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [38,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [39,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [40,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6732021 0.1616555 0.8145789 0.5456479 0.8261831 0.7681253 0.5543593
#>  [2,] 0.7329299 0.2461161 0.8411130 0.6398016 0.8577844 0.8295312 0.6956060
#>  [3,] 0.7831406 0.2746782 0.8844359 0.6242273 0.8501227 0.8146069 0.7819381
#>  [4,] 0.6494095 0.2127440 0.7131089 0.3895546 0.6644215 0.5729216 0.6777364
#>  [5,] 0.7192479 0.2510067 0.8000059 0.5243942 0.7869216 0.7162111 0.7321625
#>  [6,] 0.7472296 0.2575188 0.8324340 0.5443459 0.8014227 0.7133901 0.7234955
#>  [7,] 0.5259641 0.1355488 0.5759839 0.3331567 0.6585696 0.5750160 0.4745905
#>  [8,] 0.5991681 0.1679810 0.6923562 0.4548373 0.7351557 0.6561532 0.5194850
#>  [9,] 0.5366246 0.1113018 0.5941033 0.3526915 0.6824859 0.5895048 0.4708940
#> [10,] 0.6673053 0.1804972 0.8199612 0.4209179 0.7409063 0.6606245 0.6814651
#> [11,] 0.7305110 0.3120078 0.8278585 0.5778648 0.7843034 0.7438112 0.7417654
#> [12,] 0.7652141 0.3368697 0.8547707 0.6157178 0.8199175 0.7818531 0.7740349
#> [13,] 0.7057951 0.1846663 0.8172430 0.5236153 0.8160474 0.7448007 0.6901753
#> [14,] 0.6861577 0.2055917 0.7606947 0.4458648 0.7509756 0.6609516 0.6726676
#> [15,] 0.6634749 0.1916279 0.7475886 0.5471467 0.8184341 0.7569547 0.5444266
#> [16,] 0.7245504 0.2351926 0.8465463 0.4508764 0.7542189 0.6689530 0.7585605
#> [17,] 0.7309456 0.3342190 0.8249909 0.6345832 0.7996561 0.7610074 0.6989234
#> [18,] 0.6283012 0.2067454 0.7292767 0.3680244 0.6297090 0.5258706 0.6544702
#> [19,] 0.6471534 0.1567307 0.7136328 0.4179041 0.7860610 0.7065278 0.6054366
#> [20,] 0.7999066 0.3878903 0.8916762 0.6761468 0.8513112 0.8209990 0.7908454
#> [21,] 0.7208119 0.3358908 0.8165539 0.6057479 0.7755325 0.7480278 0.7275148
#> [22,] 0.7406926 0.2836948 0.8486274 0.5159352 0.7769684 0.7097214 0.7719060
#> [23,] 0.6284817 0.2170179 0.7233471 0.3551576 0.5987007 0.5022471 0.6652391
#> [24,] 0.7343909 0.2667040 0.8245231 0.6177844 0.8382520 0.7843125 0.6590132
#> [25,] 0.6211726 0.2147516 0.6790394 0.5638028 0.7814446 0.7303252 0.5071926
#> [26,] 0.7103781 0.1886161 0.8264143 0.5107515 0.8034017 0.7545380 0.7080105
#> [27,] 0.6820844 0.1752575 0.8240281 0.4674255 0.7443274 0.6870303 0.6727746
#> [28,] 0.7203055 0.3155072 0.8293873 0.5498931 0.7321553 0.6984627 0.7098245
#> [29,] 0.6771793 0.1727109 0.8163309 0.4577624 0.7787864 0.7060375 0.5930689
#> [30,] 0.6576671 0.2780892 0.7800715 0.4875867 0.6995309 0.6557415 0.6733718
#> [31,] 0.7183666 0.1687049 0.8184006 0.4278885 0.7922164 0.6733932 0.6916588
#> [32,] 0.5211993 0.1245482 0.5966387 0.3479697 0.6758689 0.5974527 0.4749564
#> [33,] 0.7048958 0.2063079 0.8454352 0.4916208 0.7820999 0.7067346 0.6998931
#> [34,] 0.7975962 0.3127801 0.8905281 0.6810826 0.8847500 0.8549472 0.7835707
#> [35,] 0.7851406 0.3191304 0.8873412 0.5580970 0.8211446 0.7600957 0.8247460
#> [36,] 0.6325277 0.1706952 0.7771775 0.4850842 0.7793713 0.7213447 0.5236517
#> [37,] 0.7490082 0.1817561 0.8494052 0.5749388 0.8789096 0.8140605 0.6794482
#> [38,] 0.7293573 0.3833438 0.8429117 0.6487924 0.7832919 0.7602833 0.7082050
#> [39,] 0.7233254 0.2836754 0.8189720 0.4956761 0.7634838 0.6907890 0.7542341
#> [40,] 0.7813709 0.2799460 0.8884584 0.5637964 0.8163602 0.7552335 0.8056264
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4373595 0.7491242 0.4401580
#>  [2,] 0.5761373 0.8299259 0.5855206
#>  [3,] 0.6750671 0.8514766 0.6522589
#>  [4,] 0.5039307 0.7331648 0.5155227
#>  [5,] 0.5683313 0.7983841 0.5589741
#>  [6,] 0.5995793 0.8129360 0.6071844
#>  [7,] 0.2973727 0.5922844 0.3145449
#>  [8,] 0.3802848 0.6670181 0.3663575
#>  [9,] 0.2735911 0.6054506 0.2960269
#> [10,] 0.5555494 0.7792277 0.5388434
#> [11,] 0.6355816 0.8176111 0.6443872
#> [12,] 0.6724953 0.8377438 0.6501011
#> [13,] 0.5479068 0.8254673 0.5468908
#> [14,] 0.4945144 0.7520821 0.4854590
#> [15,] 0.3812032 0.7061473 0.3888067
#> [16,] 0.6424628 0.8287222 0.6405021
#> [17,] 0.5973725 0.7884936 0.6029243
#> [18,] 0.5150457 0.7100375 0.5058894
#> [19,] 0.4319676 0.7283528 0.4541726
#> [20,] 0.7430223 0.8605284 0.7323798
#> [21,] 0.6541248 0.7953956 0.6339950
#> [22,] 0.6755214 0.8527431 0.6872320
#> [23,] 0.5318503 0.7145824 0.5337925
#> [24,] 0.5298343 0.7947154 0.5515504
#> [25,] 0.3462993 0.6625331 0.3545808
#> [26,] 0.5886319 0.8075112 0.5720895
#> [27,] 0.5706196 0.7748209 0.5507378
#> [28,] 0.6557240 0.7941034 0.6161957
#> [29,] 0.4718837 0.7678498 0.4969412
#> [30,] 0.5628526 0.7261169 0.5832947
#> [31,] 0.5639264 0.8019994 0.5150865
#> [32,] 0.2796465 0.5913686 0.3003620
#> [33,] 0.5770843 0.8066624 0.5573183
#> [34,] 0.6874394 0.8707224 0.6671081
#> [35,] 0.7590285 0.8793870 0.7413505
#> [36,] 0.4329687 0.7335439 0.4546280
#> [37,] 0.5136156 0.8197296 0.4891479
#> [38,] 0.6704449 0.7976282 0.6323406
#> [39,] 0.6315631 0.8083902 0.6406874
#> [40,] 0.7265857 0.8668768 0.6985556
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.7177726 0.2822274
#>  [2,] 0.7298413 0.2701587
#>  [3,] 0.7377216 0.2622784
#>  [4,] 0.7057805 0.2942195
#>  [5,] 0.7223066 0.2776934
#>  [6,] 0.7273682 0.2726318
#>  [7,] 0.6789318 0.3210682
#>  [8,] 0.6973072 0.3026928
#>  [9,] 0.6815043 0.3184957
#> [10,] 0.7199560 0.2800440
#> [11,] 0.7268036 0.2731964
#> [12,] 0.7322098 0.2677902
#> [13,] 0.7255449 0.2744551
#> [14,] 0.7136645 0.2863355
#> [15,] 0.7085166 0.2914834
#> [16,] 0.7287157 0.2712843
#> [17,] 0.7236944 0.2763056
#> [18,] 0.7033642 0.2966358
#> [19,] 0.7089252 0.2910748
#> [20,] 0.7392331 0.2607669
#> [21,] 0.7241123 0.2758877
#> [22,] 0.7316152 0.2683848
#> [23,] 0.7029310 0.2970690
#> [24,] 0.7246332 0.2753668
#> [25,] 0.6970255 0.3029745
#> [26,] 0.7268840 0.2731160
#> [27,] 0.7215922 0.2784078
#> [28,] 0.7238100 0.2761900
#> [29,] 0.7190160 0.2809840
#> [30,] 0.7123967 0.2876033
#> [31,] 0.7243409 0.2756591
#> [32,] 0.6802520 0.3197480
#> [33,] 0.7256452 0.2743548
#> [34,] 0.7400657 0.2599343
#> [35,] 0.7400819 0.2599181
#> [36,] 0.7115547 0.2884453
#> [37,] 0.7299291 0.2700709
#> [38,] 0.7254868 0.2745132
#> [39,] 0.7252395 0.2747605
#> [40,] 0.7387582 0.2612418
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]       [,2]      [,3]        [,4]      [,5]        [,6]
#>  [1,] 0.72270364 -1.6459617 1.4800420  0.18310148 1.5588142 1.197755576
#>  [2,] 1.00953886 -1.1194351 1.6665331  0.57450294 1.7970086 1.582308580
#>  [3,] 1.28406319 -0.9710152 2.0351243  0.50753031 1.7355634 1.480227446
#>  [4,] 0.61644448 -1.3084637 0.9105315 -0.44918493 0.6830607 0.293781171
#>  [5,] 0.94073380 -1.0932505 1.3863312  0.09765442 1.3064684 0.925744353
#>  [6,] 1.08389083 -1.0589049 1.6029769  0.17785106 1.3952101 0.911906262
#>  [7,] 0.10394985 -1.8527634 0.3063081 -0.69394215 0.6569263 0.302346280
#>  [8,] 0.40200022 -1.6000047 0.8111581 -0.18114439 1.0209402 0.646197878
#>  [9,] 0.14676127 -2.0775126 0.3809547 -0.60722940 0.7652195 0.361918830
#> [10,] 0.69602261 -1.5129823 1.5160846 -0.31900636 1.0506844 0.666078454
#> [11,] 0.99721681 -0.7907491 1.5705253  0.31401425 1.2909229 1.065872706
#> [12,] 1.18148148 -0.6772751 1.7725192  0.47141137 1.5157888 1.276498393
#> [13,] 0.87504843 -1.4850470 1.4977790  0.09453166 1.4897941 1.071072081
#> [14,] 0.78221678 -1.3517056 1.1564917 -0.21739299 1.1038222 0.667537851
#> [15,] 0.67881810 -1.4394671 1.0857925  0.18914861 1.5057744 1.136055328
#> [16,] 0.96714689 -1.1792191 1.7077661 -0.19713038 1.1212414 0.703453282
#> [17,] 0.99942559 -0.6891642 1.5505344  0.55192978 1.3841462 1.158210387
#> [18,] 0.52493589 -1.3446564 0.9909559 -0.54070144 0.5309687 0.103574982
#> [19,] 0.60654981 -1.6827572 0.9130935 -0.33138318 1.3013437 0.878579867
#> [20,] 1.38571064 -0.4561891 2.1079785  0.73611990 1.7449222 1.523130513
#> [21,] 0.94849270 -0.6816606 1.4931724  0.42947314 1.2398188 1.088121505
#> [22,] 1.04957130 -0.9262072 1.7238755  0.06376254 1.2480860 0.894031287
#> [23,] 0.52570871 -1.2831299 0.9611258 -0.59644450 0.4000544 0.008988569
#> [24,] 1.01701595 -1.0114100 1.5472978  0.48015466 1.6452792 1.290977085
#> [25,] 0.49452833 -1.2965183 0.7493606  0.25661029 1.2741045 0.996273275
#> [26,] 0.89722084 -1.4590276 1.5604250  0.04301267 1.4076924 1.122963578
#> [27,] 0.76336749 -1.5488146 1.5438804 -0.13048262 1.0685833 0.786271826
#> [28,] 0.94597747 -0.7744967 1.5812914  0.20023868 1.0055854 0.839988202
#> [29,] 0.74083911 -1.5665351 1.4916838 -0.16935381 1.2586082 0.876216087
#> [30,] 0.65291511 -0.9539595 1.2660831 -0.04966343 0.8450650 0.644373830
#> [31,] 0.93637381 -1.5948337 1.5055487 -0.29047145 1.3383372 0.723572375
#> [32,] 0.08484799 -1.9500474 0.3914789 -0.62797541 0.7348509 0.394862621
#> [33,] 0.87072144 -1.3473260 1.6992382 -0.03352007 1.2779458 0.879577008
#> [34,] 1.37133774 -0.7871539 2.0961464  0.75875143 2.0382013 1.773942161
#> [35,] 1.29587880 -0.7577711 2.0638654  0.23344236 1.5241215 1.153204281
#> [36,] 0.54307623 -1.5807083 1.2492934 -0.05968109 1.2620065 0.951141398
#> [37,] 1.09332953 -1.5044949 1.7299437  0.30203061 1.9821445 1.476613518
#> [38,] 0.99136434 -0.4753796 1.6800544  0.61373536 1.2849543 1.154233255
#> [39,] 0.96101713 -0.9263030 1.5093980 -0.01729619 1.1718752 0.803810505
#> [40,] 1.27367326 -0.9447296 2.0750904  0.25658407 1.4918795 1.126722101
#>              [,7]        [,8]      [,9]       [,10]
#>  [1,]  0.21829983 -0.25188534 1.0939469 -0.24052081
#>  [2,]  0.82646071  0.30693648 1.5851020  0.34547806
#>  [3,]  1.27699650  0.73119329 1.7462292  0.62898330
#>  [4,]  0.74338855  0.01572300 1.0107394  0.06211067
#>  [5,]  1.00562187  0.27504613 1.3762257  0.23699952
#>  [6,]  0.96186699  0.40371255 1.4692015  0.43549221
#>  [7,] -0.10172571 -0.85984049 0.3734169 -0.77895630
#>  [8,]  0.07797961 -0.48833966 0.6947290 -0.54787545
#>  [9,] -0.11655588 -0.97647837 0.4282287 -0.86628991
#> [10,]  0.76051284  0.22311877 1.2611716  0.15568712
#> [11,]  1.05516456  0.55623776 1.5002458  0.59445732
#> [12,]  1.23123624  0.71949255 1.6415360  0.61948381
#> [13,]  0.80093875  0.19221693 1.5538374  0.18811600
#> [14,]  0.72027526 -0.02194341 1.1097477 -0.05818045
#> [15,]  0.17817640 -0.48444435 0.8767451 -0.45233077
#> [16,]  1.14480377  0.58606936 1.5765982  0.57754407
#> [17,]  0.84217664  0.39452899 1.3158688  0.41766456
#> [18,]  0.63874695  0.06020108 0.8955662  0.02355875
#> [19,]  0.42816985 -0.27382789 0.9862811 -0.18382567
#> [20,]  1.33002885  1.06173653 1.8196857  1.00673064
#> [21,]  0.98204991  0.63722007 1.3577613  0.54939423
#> [22,]  1.21910477  0.73326507 1.7562796  0.78721020
#> [23,]  0.68672989  0.12757399 0.9177451  0.13537657
#> [24,]  0.65889972  0.11947896 1.3535866  0.20693686
#> [25,]  0.02877220 -0.63534602 0.6746029 -0.59896381
#> [26,]  0.88574110  0.35831244 1.4339188  0.29038149
#> [27,]  0.72076125  0.28437950 1.2357359  0.20365196
#> [28,]  0.89453178  0.64429648 1.3498396  0.47343180
#> [29,]  0.37666686 -0.11258385 1.1962095 -0.01223545
#> [30,]  0.72347496  0.25274725 0.9750094  0.33631323
#> [31,]  0.80788561  0.25711271 1.3988380  0.06036452
#> [32,] -0.10025835 -0.94621578 0.3696260 -0.84557463
#> [33,]  0.84678887  0.31081569 1.4284673  0.23028571
#> [34,]  1.28659721  0.78817525 1.9073608  0.69513433
#> [35,]  1.54883917  1.14736092 1.9866376  1.05299949
#> [36,]  0.09467759 -0.26974907 1.0126780 -0.18198864
#> [37,]  0.75123708  0.05447586 1.5145166 -0.04341520
#> [38,]  0.88668223  0.71019818 1.3715358  0.54227130
#> [39,]  1.12132360  0.53892838 1.4395838  0.57834908
#> [40,]  1.42183802  0.97736833 1.8736218  0.84042930
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.5319093 -0.4015303
#>  [2,] 0.5558162 -0.4380015
#>  [3,] 0.5702775 -0.4638826
#>  [4,] 0.4830988 -0.3918794
#>  [5,] 0.5307934 -0.4251385
#>  [6,] 0.5426768 -0.4386341
#>  [7,] 0.4320661 -0.3168008
#>  [8,] 0.4790196 -0.3554881
#>  [9,] 0.4408549 -0.3198386
#> [10,] 0.5195190 -0.4247244
#> [11,] 0.5398772 -0.4385882
#> [12,] 0.5555397 -0.4503235
#> [13,] 0.5388338 -0.4333017
#> [14,] 0.5102185 -0.4030303
#> [15,] 0.5136805 -0.3745100
#> [16,] 0.5365791 -0.4515370
#> [17,] 0.5396416 -0.4232199
#> [18,] 0.4758198 -0.3875501
#> [19,] 0.5008407 -0.3893290
#> [20,] 0.5730455 -0.4689406
#> [21,] 0.5343190 -0.4306334
#> [22,] 0.5446058 -0.4582271
#> [23,] 0.4714039 -0.3898905
#> [24,] 0.5458890 -0.4216725
#> [25,] 0.4876424 -0.3455309
#> [26,] 0.5400969 -0.4387732
#> [27,] 0.5250912 -0.4272822
#> [28,] 0.5304357 -0.4330041
#> [29,] 0.5260205 -0.4135655
#> [30,] 0.5036071 -0.4034460
#> [31,] 0.5322984 -0.4337991
#> [32,] 0.4379697 -0.3169603
#> [33,] 0.5371431 -0.4354957
#> [34,] 0.5791220 -0.4671881
#> [35,] 0.5655972 -0.4807971
#> [36,] 0.5113714 -0.3915756
#> [37,] 0.5598989 -0.4343640
#> [38,] 0.5408978 -0.4309455
#> [39,] 0.5315491 -0.4390532
#> [40,] 0.5647657 -0.4747588
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.2822274 -0.2822274
#>  [2,] -0.7298413  0.7298413
#>  [3,]  0.2622784 -0.2622784
#>  [4,]  0.2942195 -0.2942195
#>  [5,] -0.7223066  0.7223066
#>  [6,]  0.2726318 -0.2726318
#>  [7,]  0.3210682 -0.3210682
#>  [8,]  0.3026928 -0.3026928
#>  [9,]  0.3184957 -0.3184957
#> [10,] -0.7199560  0.7199560
#> [11,] -0.7268036  0.7268036
#> [12,] -0.7322098  0.7322098
#> [13,] -0.7255449  0.7255449
#> [14,] -0.7136645  0.7136645
#> [15,] -0.7085166  0.7085166
#> [16,] -0.7287157  0.7287157
#> [17,] -0.7236944  0.7236944
#> [18,] -0.7033642  0.7033642
#> [19,] -0.7089252  0.7089252
#> [20,] -0.7392331  0.7392331
#> [21,] -0.7241123  0.7241123
#> [22,]  0.2683848 -0.2683848
#> [23,]  0.2970690 -0.2970690
#> [24,] -0.7246332  0.7246332
#> [25,] -0.6970255  0.6970255
#> [26,]  0.2731160 -0.2731160
#> [27,]  0.2784078 -0.2784078
#> [28,] -0.7238100  0.7238100
#> [29,]  0.2809840 -0.2809840
#> [30,]  0.2876033 -0.2876033
#> [31,] -0.7243409  0.7243409
#> [32,]  0.3197480 -0.3197480
#> [33,] -0.7256452  0.7256452
#> [34,] -0.7400657  0.7400657
#> [35,] -0.7400819  0.7400819
#> [36,]  0.2884453 -0.2884453
#> [37,] -0.7299291  0.7299291
#> [38,] -0.7254868  0.7254868
#> [39,]  0.2747605 -0.2747605
#> [40,]  0.2612418 -0.2612418
#> 
#> $L
#> [1] 0.7050730 0.7072429 0.6925005 0.7061323 0.7923307 0.8834491
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5507246