Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.3474969 -0.10197312 -0.12698376 -0.17147605 -0.050364772 -0.11242548
#>  [2,] -0.2624238 -0.15173391 -0.01102919 -0.12971114 -0.103532578  0.05323078
#>  [3,] -0.2706203 -0.19319980 -0.08396024 -0.03241108 -0.020657727 -0.03163077
#>  [4,] -0.3117949 -0.15649905 -0.11578415  0.02831104 -0.063490245 -0.16552740
#>  [5,] -0.3239281 -0.07538987 -0.09186228 -0.02215473 -0.009894987 -0.11904551
#>  [6,] -0.2324292  0.06093504 -0.07942463 -0.08870754 -0.010243499 -0.05484979
#>  [7,] -0.2267377 -0.19524032 -0.03440277 -0.08421395 -0.192117428 -0.03181970
#>  [8,] -0.2864628  0.01913959 -0.05205400 -0.05088299  0.070968081  0.04406858
#>  [9,] -0.3773450 -0.08100891 -0.08275910 -0.05115204 -0.124858656  0.01519924
#> [10,] -0.3497068 -0.19695543 -0.07924420 -0.14542858 -0.003918785 -0.03136735
#>                V15         V16          V17         V18         V19         V2
#>  [1,] -0.081529648 -0.10221852 -0.138485810 -0.15065305 -0.02715939 -0.2647186
#>  [2,] -0.051931507 -0.04639094  0.038074507 -0.05389162 -0.01472897 -0.2627107
#>  [3,] -0.048661417 -0.13389780 -0.113643674 -0.07292803  0.05416465 -0.3827515
#>  [4,] -0.072944625  0.02606622  0.067829441  0.08805300 -0.05981526 -0.1913467
#>  [5,] -0.073727023  0.04023073  0.044626168  0.08095307  0.13123505 -0.4377823
#>  [6,] -0.098605074 -0.06130478 -0.085955497  0.04079610 -0.09044637 -0.2053980
#>  [7,]  0.006913608 -0.03304272  0.003531386  0.08307347  0.02331600 -0.2406517
#>  [8,] -0.032994794  0.04118964 -0.089076511  0.01965305  0.08382088 -0.1891353
#>  [9,] -0.120284502 -0.14213021 -0.089560692 -0.14357339  0.04274125 -0.2187797
#> [10,]  0.023154654 -0.09621167  0.007336227 -0.03139099  0.05017084 -0.4047942
#>                 V20         V21        V22         V23        V24          V25
#>  [1,]  0.0669570633 -0.02296095 0.02851459  0.11216017 0.04365048  0.035797438
#>  [2,] -0.0002754081  0.16022808 0.03947008  0.09549446 0.10061665  0.199554391
#>  [3,]  0.0326402491 -0.02347922 0.11598459  0.09996280 0.09676203 -0.009551403
#>  [4,]  0.1630316718  0.12527969 0.12397351  0.10430619 0.23492811  0.250462886
#>  [5,]  0.1203756383  0.25378155 0.22062217  0.31477916 0.21913662  0.269910539
#>  [6,]  0.0168211075  0.03453700 0.04982871  0.06026720 0.12695629  0.143372423
#>  [7,]  0.0827885030 -0.04858753 0.04553880 -0.05206093 0.06141723  0.078026731
#>  [8,]  0.0764908377  0.07253026 0.11249636  0.24527291 0.21194419  0.235802102
#>  [9,] -0.1037291286  0.04637170 0.12605039  0.07718873 0.03812249  0.145999022
#> [10,] -0.0507134206  0.10269887 0.09449243  0.12019713 0.05977038  0.138641121
#>              V26         V27        V28         V29         V3          V30
#>  [1,] 0.09647166  0.11101471 0.10764211 -0.03017814 -0.1740532 -0.065866687
#>  [2,] 0.22791501  0.28662532 0.13999061  0.04877511 -0.2823178  0.151009323
#>  [3,] 0.08144992  0.12201358 0.11367522  0.07572996 -0.3209343  0.003983801
#>  [4,] 0.14052409  0.07547921 0.07841674  0.10380273 -0.1744710 -0.018947982
#>  [5,] 0.29385967  0.15446993 0.18829858  0.16724616 -0.3975670  0.120753756
#>  [6,] 0.22373700  0.17404631 0.36097765  0.13797398 -0.2267413  0.162018628
#>  [7,] 0.05657669 -0.04207142 0.00593181  0.06772028 -0.1578517 -0.038278095
#>  [8,] 0.26702334  0.26181376 0.18137334  0.10478071 -0.2606879  0.138859387
#>  [9,] 0.06353492  0.18182603 0.10332621  0.16307641 -0.2538158 -0.058443940
#> [10,] 0.12918401  0.19938537 0.04986607  0.09920285 -0.3597443  0.084399145
#>                V31         V32          V33          V34          V35
#>  [1,] -0.058117841 -0.13039044 -0.069195228 -0.084408448 -0.129563610
#>  [2,]  0.007636607  0.05862618  0.101750653 -0.006348573 -0.022707913
#>  [3,] -0.080978327  0.00577495  0.034049284 -0.013378170 -0.062949053
#>  [4,]  0.061920908  0.08649680  0.100495739 -0.091612951  0.052735139
#>  [5,]  0.154381283  0.13192061  0.121501355  0.093596183  0.040413197
#>  [6,]  0.064354014  0.07043028 -0.013776887  0.064228769  0.005907915
#>  [7,] -0.093975230 -0.24094592 -0.193097084 -0.163291497 -0.151527073
#>  [8,]  0.020605026  0.03840278 -0.004332552 -0.053881814  0.058394910
#>  [9,]  0.087644001 -0.03802696 -0.089019882 -0.078473001 -0.005557053
#> [10,]  0.117710285  0.07908661  0.044123112  0.061151138  0.060968500
#>                 V36         V37          V38          V39         V4
#>  [1,]  1.495406e-02 -0.06633742 -0.036707693  0.007864916 -0.2584536
#>  [2,]  4.339094e-02  0.04656404  0.006745415  0.089724761 -0.2025143
#>  [3,] -8.564285e-02 -0.15227721 -0.014266533 -0.094859741 -0.1822839
#>  [4,] -2.621156e-03 -0.07839830  0.003098953 -0.013996464 -0.2418602
#>  [5,]  1.312857e-01  0.03503229 -0.052901671 -0.047677119 -0.2955428
#>  [6,]  1.386488e-01  0.12010988  0.036154196  0.014565589 -0.2715379
#>  [7,] -1.071563e-01 -0.06391397 -0.206296590 -0.045369473 -0.1852376
#>  [8,]  8.383533e-05  0.02296634 -0.045187120  0.103727740 -0.2321703
#>  [9,] -9.997651e-04  0.00230518 -0.081363672 -0.133718668 -0.2264214
#> [10,] -9.867336e-02  0.01440706 -0.064480546 -0.135352514 -0.3224515
#>                V40         V41          V42          V43         V44
#>  [1,] -0.007313255 -0.20203981 -0.184790059 -0.207104785 -0.11439013
#>  [2,]  0.006431318 -0.09892034 -0.120499850 -0.071267342 -0.02546111
#>  [3,] -0.000518266 -0.05996953 -0.021041992 -0.007792362 -0.16329424
#>  [4,] -0.044509032 -0.05363127 -0.005767999 -0.155020579 -0.12589067
#>  [5,] -0.064714535  0.03011568 -0.122162799 -0.041304849 -0.10800842
#>  [6,] -0.003946195 -0.04157834 -0.046579242 -0.035647353 -0.05668118
#>  [7,] -0.140951759 -0.18990733 -0.087509046 -0.124730716 -0.20155874
#>  [8,]  0.032043742 -0.11330331 -0.116160197 -0.097296447 -0.10817506
#>  [9,]  0.056337003 -0.07589638 -0.014422044 -0.064167856 -0.07331866
#> [10,] -0.064031774 -0.04952629  0.036240507 -0.106328494 -0.09089892
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.24931687 -0.08591248 -0.1453742 -0.2366581 -0.1592906 -0.1825239
#>  [2,] -0.03311114 -0.11868793 -0.1718325 -0.1922802 -0.2501835 -0.2314555
#>  [3,] -0.24649843 -0.24714491 -0.2577157 -0.1926361 -0.3332532 -0.2575724
#>  [4,] -0.02296360 -0.20066719 -0.1302314 -0.2516254 -0.2491549 -0.2729671
#>  [5,] -0.08212975 -0.26647426 -0.1587493 -0.2369211 -0.3438058 -0.2640268
#>  [6,] -0.00577983 -0.04618612 -0.1753877 -0.1364485 -0.3043886 -0.1805313
#>  [7,] -0.24583040 -0.27950453 -0.2824907 -0.3208854 -0.1652406 -0.1666379
#>  [8,] -0.13190878 -0.12293363 -0.1295903 -0.1113503 -0.2641560 -0.1293773
#>  [9,] -0.13364047 -0.08577150 -0.2187580 -0.2723263 -0.2202343 -0.2984109
#> [10,] -0.20785283 -0.13405195 -0.2701385 -0.3275140 -0.2629805 -0.3267323
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2013848 -0.2938952 -0.3147176 -0.2422256 -0.3464877 -0.3157319
#>  [2,] -0.3328004 -0.3255309 -0.3311637 -0.3326276 -0.2745451 -0.3479148
#>  [3,] -0.2543263 -0.3218019 -0.2855671 -0.3201127 -0.3697567 -0.3157500
#>  [4,] -0.2167286 -0.2524971 -0.3196246 -0.2086333 -0.2230923 -0.1770816
#>  [5,] -0.3698568 -0.3854440 -0.4228720 -0.3808544 -0.3799066 -0.4710810
#>  [6,] -0.2653465 -0.1885929 -0.3164512 -0.3424484 -0.3666562 -0.2938575
#>  [7,] -0.3637498 -0.3919622 -0.2067103 -0.2014493 -0.3438533 -0.2149397
#>  [8,] -0.3585363 -0.2445154 -0.2513926 -0.3761567 -0.2691382 -0.2278532
#>  [9,] -0.2821388 -0.3926342 -0.3243624 -0.3502272 -0.2279454 -0.3249714
#> [10,] -0.3912396 -0.4029858 -0.3340175 -0.4433882 -0.4314491 -0.3355139
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.1809622 -0.3433416 -0.3774470 -0.3556523 -0.20799228 -0.3122756
#>  [2,] -0.2913496 -0.2292038 -0.3885415 -0.2798049 -0.23471668 -0.3929319
#>  [3,] -0.3437118 -0.4109889 -0.2502544 -0.3783880 -0.10299503 -0.2925115
#>  [4,] -0.2672414 -0.2382145 -0.3196083 -0.3217321 -0.22333503 -0.1898029
#>  [5,] -0.3604051 -0.4274869 -0.3615598 -0.3623161 -0.30232697 -0.3714331
#>  [6,] -0.3424976 -0.3472332 -0.3369256 -0.2614620 -0.13498993 -0.3103691
#>  [7,] -0.2951345 -0.2449577 -0.3018911 -0.2653467 -0.26304567 -0.2926416
#>  [8,] -0.3357842 -0.4007568 -0.2495024 -0.3551729 -0.08340057 -0.3071928
#>  [9,] -0.2517390 -0.2618386 -0.3766662 -0.3019417 -0.16407422 -0.3579350
#> [10,] -0.4266912 -0.4533652 -0.3140918 -0.4396102 -0.12678698 -0.3914422
#>                V7          V8          V9
#>  [1,] -0.12536272 -0.19507187 -0.14080198
#>  [2,] -0.08187204 -0.20454601 -0.15099373
#>  [3,] -0.14612678 -0.19842259 -0.15045683
#>  [4,] -0.20805246 -0.10563957 -0.13097061
#>  [5,] -0.23557725 -0.20072713 -0.07135776
#>  [6,] -0.21996244 -0.02576513 -0.12778098
#>  [7,] -0.25454400 -0.11005586 -0.10148902
#>  [8,] -0.21996113 -0.06007799 -0.03194827
#>  [9,] -0.12144853 -0.23395186 -0.20230021
#> [10,] -0.14892363 -0.13812443 -0.17570327
#> 
#> $W[[2]]
#>              [,1]       [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.008205381 -0.1038788 -0.11854556 -0.04617716 -0.04878265 -0.1461631
#> [2,]  0.126323604  0.1836036  0.07856653  0.13277959  0.04911890  0.1443331
#>             [,7]        [,8]        [,9]      [,10]
#> [1,] -0.17476235 -0.14226092 -0.01886651 -0.1339703
#> [2,]  0.03640784  0.07358466  0.20077778  0.0491877
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -3.129491e-07 -6.501756e-06 -4.142803e-06 -2.023605e-05 -2.634723e-05
#>  [2,]  4.705872e-05  2.835290e-04  3.713876e-04  3.349500e-04  2.857192e-04
#>  [3,]  2.253322e-05  2.075694e-04  2.478436e-04  3.533600e-04  3.350980e-04
#>  [4,]  1.354412e-05  2.518072e-04  2.712078e-04  6.412331e-04  6.512082e-04
#>  [5,] -3.517445e-05  2.670066e-05  1.236919e-05  5.774805e-04  5.749621e-04
#>  [6,]  4.717128e-05  2.676873e-04  3.587035e-04  3.356623e-04  2.919866e-04
#>  [7,]  5.899678e-06 -1.347118e-04 -4.807081e-05 -3.361716e-04 -4.549057e-04
#>  [8,]  2.356913e-05  2.697261e-04  3.219520e-04  5.927035e-04  5.727922e-04
#>  [9,]  2.740151e-05  1.799618e-04  2.449103e-04  1.940631e-04  1.407982e-04
#> [10,]  1.007254e-06  1.710740e-04  1.614018e-04  5.208211e-04  5.475768e-04
#>                 V14           V15           V16           V17           V18
#>  [1,] -4.767872e-05 -6.473338e-05 -9.310202e-05 -1.022246e-04 -1.038733e-04
#>  [2,]  1.853371e-04  6.428389e-05 -4.449440e-05 -5.205298e-05 -8.442359e-05
#>  [3,]  3.441100e-04  3.374391e-04  4.138786e-04  4.858618e-04  4.867138e-04
#>  [4,]  8.175088e-04  9.481382e-04  1.309677e-03  1.532748e-03  1.562157e-03
#>  [5,]  7.418287e-04  8.573005e-04  1.283830e-03  1.528614e-03  1.587428e-03
#>  [6,]  2.008916e-04  8.741254e-05 -1.351970e-05 -1.156369e-05 -3.733256e-05
#>  [7,] -8.023875e-04 -1.120352e-03 -1.602869e-03 -1.761168e-03 -1.758514e-03
#>  [8,]  6.374011e-04  6.571549e-04  8.612775e-04  1.019704e-03  1.028998e-03
#>  [9,]  1.966930e-05 -1.051236e-04 -2.369695e-04 -2.651811e-04 -2.830090e-04
#> [10,]  7.423112e-04  9.131042e-04  1.300575e-03  1.523778e-03  1.555231e-03
#>                 V19            V2           V20           V21           V22
#>  [1,] -8.323024e-05 -1.566499e-06 -7.298388e-05 -8.537164e-05 -0.0001061447
#>  [2,]  1.259538e-04  5.995276e-05  3.247719e-04  3.224511e-04  0.0002857907
#>  [3,]  4.846222e-04  3.588537e-05  5.071573e-04  5.417452e-04  0.0006794175
#>  [4,]  1.303404e-03  3.802683e-05  1.106695e-03  1.188382e-03  0.0016563805
#>  [5,]  1.218880e-03 -2.528072e-05  7.333736e-04  7.199780e-04  0.0014198762
#>  [6,]  1.614711e-04  5.909179e-05  3.417601e-04  3.399512e-04  0.0003118003
#>  [7,] -1.349091e-03 -1.254151e-05 -1.222194e-03 -1.431935e-03 -0.0018396697
#>  [8,]  9.337047e-04  4.372772e-05  8.422709e-04  8.815865e-04  0.0012287711
#>  [9,] -9.476948e-05  3.259166e-05  5.214398e-05  2.654300e-05 -0.0000377012
#> [10,]  1.238458e-03  2.220166e-05  9.961036e-04  1.078421e-03  0.0015351659
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0001304095 -0.0001289978 -1.312840e-04 -0.0001265922 -1.123022e-04
#>  [2,]  0.0002379959  0.0001784666  9.158561e-05  0.0001750240  3.305179e-04
#>  [3,]  0.0007787608  0.0006844985  6.482554e-04  0.0006776656  6.989877e-04
#>  [4,]  0.0020481147  0.0018170617  1.793884e-03  0.0017717258  1.666504e-03
#>  [5,]  0.0020114951  0.0016909168  1.642330e-03  0.0015522641  1.437886e-03
#>  [6,]  0.0002640977  0.0001944393  9.670900e-05  0.0001653252  3.071861e-04
#>  [7,] -0.0022666345 -0.0021922568 -2.280508e-03 -0.0022445881 -1.932289e-03
#>  [8,]  0.0015069361  0.0012997799  1.236230e-03  0.0012475031  1.260514e-03
#>  [9,] -0.0001202316 -0.0001483705 -2.102224e-04 -0.0001507029 -1.067084e-05
#> [10,]  0.0019197735  0.0017067277  1.709231e-03  0.0016661671  1.512272e-03
#>                 V28           V29            V3           V30           V31
#>  [1,] -1.158424e-04 -0.0001326305 -2.650045e-06 -1.269275e-04 -1.281549e-04
#>  [2,]  3.382037e-04  0.0001694792  6.544740e-05  7.457600e-05 -1.029183e-04
#>  [3,]  6.895606e-04  0.0006713144  4.889869e-05  5.909373e-04  5.590982e-04
#>  [4,]  1.637969e-03  0.0017902097  7.324195e-05  1.664814e-03  1.807693e-03
#>  [5,]  1.363496e-03  0.0016138003  1.506066e-05  1.414515e-03  1.875906e-03
#>  [6,]  3.179886e-04  0.0001649225  6.585132e-05  7.298526e-05 -9.520651e-05
#>  [7,] -1.911750e-03 -0.0021933193 -3.482009e-05 -2.127852e-03 -2.189208e-03
#>  [8,]  1.243765e-03  0.0012841528  7.164901e-05  1.134280e-03  1.215629e-03
#>  [9,] -5.542760e-07 -0.0001423120  3.289465e-05 -2.101101e-04 -3.322224e-04
#> [10,]  1.467137e-03  0.0016423721  5.546291e-05  1.561467e-03  1.759765e-03
#>                 V32           V33           V34           V35           V36
#>  [1,] -1.024684e-04 -9.761037e-05 -0.0001083407 -0.0001139098 -0.0001149648
#>  [2,] -3.556655e-05 -2.980732e-05 -0.0001506681 -0.0003148688 -0.0004104326
#>  [3,]  4.962433e-04  4.558141e-04  0.0004080451  0.0003206233  0.0003293463
#>  [4,]  1.515131e-03  1.372694e-03  0.0013839456  0.0013316548  0.0014552111
#>  [5,]  1.525284e-03  1.253858e-03  0.0012460048  0.0012105022  0.0015189299
#>  [6,] -3.418133e-05 -2.146370e-05 -0.0001329579 -0.0002843770 -0.0003779231
#>  [7,] -1.795908e-03 -1.712707e-03 -0.0018710120 -0.0019224418 -0.0020633600
#>  [8,]  1.039704e-03  9.109672e-04  0.0008537537  0.0007409133  0.0008166115
#>  [9,] -2.391485e-04 -2.338102e-04 -0.0003404559 -0.0004695795 -0.0005260981
#> [10,]  1.470162e-03  1.331066e-03  0.0013700869  0.0013645345  0.0015109389
#>                 V37           V38           V39            V4           V40
#>  [1,] -9.853774e-05 -7.475197e-05 -6.012065e-05 -2.781123e-06 -6.020566e-05
#>  [2,] -2.802324e-04 -7.447008e-05  6.146734e-06  8.754879e-05 -7.480146e-05
#>  [3,]  3.749573e-04  3.605560e-04  3.343628e-04  5.781300e-05  3.048621e-04
#>  [4,]  1.425012e-03  1.139524e-03  9.660166e-04  7.386587e-05  9.758775e-04
#>  [5,]  1.550868e-03  1.178436e-03  1.023188e-03 -9.545300e-06  1.133217e-03
#>  [6,] -2.534011e-04 -5.067276e-05  2.856662e-05  8.671933e-05 -5.265059e-05
#>  [7,] -1.773933e-03 -1.261578e-03 -9.945267e-04 -3.664785e-05 -1.021218e-03
#>  [8,]  8.708095e-04  7.521016e-04  6.673069e-04  7.886977e-05  6.621701e-04
#>  [9,] -3.981313e-04 -2.043655e-04 -1.119446e-04  4.609373e-05 -1.613104e-04
#> [10,]  1.456391e-03  1.119704e-03  9.199194e-04  5.268018e-05  9.526630e-04
#>                 V41           V42           V43           V44           V45
#>  [1,] -5.218304e-05 -2.869995e-05 -0.0000193041 -6.149713e-06  9.979245e-06
#>  [2,]  1.401187e-05  2.396018e-04  0.0002627362  2.651115e-04  3.909377e-04
#>  [3,]  3.206439e-04  3.149109e-04  0.0002490052  1.925276e-04  1.727022e-04
#>  [4,]  9.389034e-04  6.540556e-04  0.0004182699  2.232292e-04  5.066741e-06
#>  [5,]  1.055624e-03  6.106389e-04  0.0003069188  8.936471e-05 -2.878871e-04
#>  [6,]  2.358411e-05  2.352911e-04  0.0002569359  2.478799e-04  3.593560e-04
#>  [7,] -9.335580e-04 -5.044443e-04 -0.0002713613 -9.102173e-05  1.678908e-04
#>  [8,]  6.808182e-04  5.687806e-04  0.0004068947  2.709093e-04  1.526956e-04
#>  [9,] -9.521961e-05  9.665068e-05  0.0001414905  1.772329e-04  2.837936e-04
#> [10,]  8.994541e-04  5.600118e-04  0.0003179449  1.318339e-04 -1.024022e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  2.453395e-06 -3.760499e-06  1.574217e-06  2.287404e-06 -6.692647e-06
#>  [2,]  2.830087e-04  1.737079e-04  1.579667e-04  9.951868e-05  8.270769e-05
#>  [3,]  1.449324e-04  1.206834e-04  7.784105e-05  4.247322e-05  8.852769e-05
#>  [4,]  6.811238e-05  1.440770e-04  3.493584e-05  2.694438e-06  1.664015e-04
#>  [5,] -1.623372e-04  3.409466e-05 -7.718536e-05 -5.695527e-05  1.200969e-04
#>  [6,]  2.597318e-04  1.598645e-04  1.456294e-04  9.213276e-05  8.375056e-05
#>  [7,]  4.020043e-05 -5.047445e-05  4.661396e-05  5.531459e-05 -1.220115e-04
#>  [8,]  1.480708e-04  1.679749e-04  8.720172e-05  4.353135e-05  1.465214e-04
#>  [9,]  1.896602e-04  1.066751e-04  1.081322e-04  7.220426e-05  3.736382e-05
#> [10,] -7.559027e-06  9.342968e-05 -8.146633e-06 -2.452837e-05  1.431965e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -2.362813e-06 -2.247044e-07 -1.265880e-06 -1.441132e-06 -1.025703e-06
#>  [2,]  1.433038e-05  3.004654e-05  1.693609e-05  7.005799e-06  1.026291e-05
#>  [3,]  1.749120e-05  1.727399e-05  1.345170e-05  8.758654e-06  8.585485e-06
#>  [4,]  3.522806e-05  1.471766e-05  2.079633e-05  1.832804e-05  1.447173e-05
#>  [5,]  2.160785e-05 -9.334514e-06  4.151791e-06  8.496816e-06  4.066048e-06
#>  [6,]  1.358563e-05  2.889145e-05  1.638449e-05  7.439119e-06  1.049646e-05
#>  [7,] -3.458540e-05 -1.960051e-06 -1.868712e-05 -1.845854e-05 -1.401507e-05
#>  [8,]  2.947547e-05  2.071824e-05  1.979816e-05  1.441107e-05  1.316778e-05
#>  [9,]  4.230175e-06  1.826201e-05  7.746685e-06  1.436950e-06  3.676614e-06
#> [10,]  2.996694e-05  7.161984e-06  1.628770e-05  1.587097e-05  1.137867e-05
#>                 V55           V56           V57           V58           V59
#>  [1,] -1.654140e-06 -1.363753e-06 -1.946250e-06 -1.105047e-06 -9.813569e-07
#>  [2,]  2.791310e-06  1.841811e-06 -3.845521e-07  4.336250e-06  6.017765e-06
#>  [3,]  5.376301e-06  5.582039e-06  7.678457e-06  5.468421e-06  6.582043e-06
#>  [4,]  1.350414e-05  1.430184e-05  2.414574e-05  1.147815e-05  1.340185e-05
#>  [5,]  5.804600e-06  7.445830e-06  1.758055e-05  2.096149e-06  6.865984e-06
#>  [6,]  3.166142e-06  2.140819e-06  2.681065e-07  4.768543e-06  6.644290e-06
#>  [7,] -1.938525e-05 -1.972238e-05 -3.144955e-05 -1.220287e-05 -1.207108e-05
#>  [8,]  9.660163e-06  8.955293e-06  1.501194e-05  8.617048e-06  1.115019e-05
#>  [9,] -1.329083e-06 -1.493857e-06 -4.542110e-06  4.030083e-07  1.527474e-06
#> [10,]  1.197793e-05  1.340661e-05  2.334584e-05  9.966750e-06  1.114349e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -1.738740e-05 -1.164451e-06 -1.965559e-05 -1.695667e-05 -4.031315e-06
#>  [2,]  2.582596e-05  1.256594e-06  2.696182e-05  8.515129e-05  2.412305e-04
#>  [3,]  1.077384e-04  5.246623e-06  1.082099e-04  1.284514e-04  1.571018e-04
#>  [4,]  2.870446e-04  1.421558e-05  2.839570e-04  2.742599e-04  1.575478e-04
#>  [5,]  2.690521e-04  8.413655e-06  2.357083e-04  1.951873e-04 -6.420881e-05
#>  [6,]  3.098105e-05  1.743927e-06  3.200553e-05  8.548548e-05  2.259307e-04
#>  [7,] -3.203333e-04 -1.730597e-05 -3.418282e-04 -2.985627e-04 -1.055855e-04
#>  [8,]  2.039452e-04  9.122569e-06  1.960103e-04  2.111457e-04  1.864193e-04
#>  [9,] -1.889404e-05 -1.974592e-06 -2.256570e-05  2.519546e-05  1.546476e-04
#> [10,]  2.739731e-04  1.364319e-05  2.701794e-04  2.421969e-04  9.258401e-05
#> 
#> $vW[[2]]
#>             [,1]       [,2]       [,3]      [,4]      [,5]       [,6]
#> [1,]  0.06421446  0.1376892  0.1072582  0.126533  0.159214  0.1399425
#> [2,] -0.06421446 -0.1376892 -0.1072582 -0.126533 -0.159214 -0.1399425
#>             [,7]       [,8]       [,9]      [,10]
#> [1,]  0.06049736  0.1434373  0.1022206  0.1277103
#> [2,] -0.06049736 -0.1434373 -0.1022206 -0.1277103
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] -0.04536446  0.63776223  0.58226496  0.27251805  0.91669885  0.60349754
#>  [7]  0.14003079  0.54146240  0.35382468  0.47755377
#> 
#> $B[[2]]
#> [1] -0.009918031  0.149949031
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -2.020537e-04  7.388049e-05  9.687946e-04  2.814426e-03  2.664319e-03
#>  [6]  1.019167e-04 -3.452412e-03  1.932730e-03 -3.762264e-04  2.690799e-03
#> 
#> $vB[[2]]
#> [1]  0.1706689 -0.1706689
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#>  [2,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#>  [3,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#>  [4,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#>  [5,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#>  [6,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#>  [7,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#>  [8,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [9,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [10,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [11,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [12,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [13,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [14,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [15,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [16,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [17,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [18,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [19,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [20,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [21,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [22,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [23,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [24,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [25,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [26,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [27,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [28,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [29,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [30,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [31,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [32,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [33,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [34,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [35,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [36,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [37,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [38,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [39,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [40,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#>  [2,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#>  [3,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#>  [4,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#>  [5,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#>  [6,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#>  [7,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#>  [8,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [9,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [10,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [11,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [12,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [13,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [14,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [15,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [16,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [17,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [18,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [19,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [20,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [21,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [22,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [23,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [24,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [25,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [26,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [27,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [28,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [29,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [30,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [31,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [32,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [33,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [34,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [35,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [36,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [37,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [38,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [39,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [40,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#>  [2,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#>  [3,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#>  [4,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#>  [5,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#>  [6,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#>  [7,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#>  [8,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [9,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [10,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [11,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [12,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [13,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [14,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [15,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [16,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [17,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [18,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [19,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [20,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [21,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [22,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [23,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [24,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [25,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [26,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [27,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [28,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [29,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [30,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [31,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [32,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [33,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [34,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [35,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [36,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [37,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [38,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [39,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [40,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#>  [2,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#>  [3,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#>  [4,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#>  [5,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#>  [6,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#>  [7,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#>  [8,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [9,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [10,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [11,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [12,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [13,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [14,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [15,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [16,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [17,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [18,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [19,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [20,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [21,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [22,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [23,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [24,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [25,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [26,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [27,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [28,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [29,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [30,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [31,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [32,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [33,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [34,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [35,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [36,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [37,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [38,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [39,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [40,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#>  [2,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#>  [3,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#>  [4,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#>  [5,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#>  [6,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#>  [7,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#>  [8,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [9,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [10,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [11,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [12,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [13,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [14,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [15,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [16,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [17,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [18,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [19,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [20,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [21,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [22,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [23,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [24,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [25,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [26,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [27,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [28,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [29,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [30,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [31,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [32,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [33,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [34,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [35,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [36,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [37,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [38,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [39,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [40,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#>  [2,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#>  [3,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#>  [4,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#>  [5,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#>  [6,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#>  [7,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#>  [8,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [9,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [10,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [11,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [12,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [13,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [14,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [15,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [16,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [17,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [18,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [19,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [20,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [21,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [22,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [23,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [24,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [25,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [26,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [27,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [28,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [29,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [30,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [31,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [32,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [33,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [34,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [35,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [36,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [37,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [38,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [39,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [40,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2507979 0.7918228 0.5037173 0.6484747 0.9083480 0.8495956 0.1654430
#>  [2,] 0.3410075 0.7518531 0.5913802 0.6832796 0.8896391 0.7788794 0.3943936
#>  [3,] 0.2331614 0.6145126 0.4534618 0.5605235 0.7863389 0.5905034 0.3433862
#>  [4,] 0.2857537 0.7241769 0.5261955 0.5816780 0.8460525 0.7887660 0.2101682
#>  [5,] 0.2417859 0.6527443 0.4859310 0.5722256 0.8203120 0.6468984 0.3199433
#>  [6,] 0.2887522 0.8388800 0.5443340 0.7342998 0.9425011 0.8709551 0.2173227
#>  [7,] 0.3865076 0.8389793 0.6391632 0.7971250 0.9547783 0.8179816 0.4034439
#>  [8,] 0.1772403 0.6718339 0.3523123 0.5119647 0.8172122 0.7493581 0.1341053
#>  [9,] 0.4051628 0.8372250 0.6667268 0.7628514 0.9417088 0.8579739 0.3526009
#> [10,] 0.2779090 0.8366851 0.5414469 0.7395524 0.9470930 0.8702248 0.2113114
#> [11,] 0.2717305 0.7242417 0.5231392 0.6582862 0.8861323 0.7275078 0.2600214
#> [12,] 0.4223631 0.7718811 0.6330113 0.6933285 0.9017353 0.7908304 0.3885633
#> [13,] 0.1889495 0.6787923 0.4290357 0.5945899 0.8743332 0.7367903 0.2093211
#> [14,] 0.4151243 0.8347038 0.6587952 0.7621301 0.9410593 0.8440042 0.3816601
#> [15,] 0.3518177 0.8225410 0.6172519 0.7236619 0.9306643 0.8464424 0.3078591
#> [16,] 0.3545008 0.8182575 0.6367150 0.7398393 0.9438944 0.8450280 0.2720355
#> [17,] 0.3722956 0.7974086 0.6389903 0.6960789 0.9045547 0.8203708 0.3405779
#> [18,] 0.2737308 0.7975657 0.5426015 0.6900065 0.9182933 0.8241258 0.2222271
#> [19,] 0.3091375 0.7869460 0.5754728 0.6838804 0.9153432 0.8133146 0.2592596
#> [20,] 0.2989174 0.7500225 0.5126621 0.6597205 0.8948326 0.7625556 0.2590693
#> [21,] 0.3467959 0.7983471 0.6090429 0.7389663 0.9264675 0.8017422 0.3640084
#> [22,] 0.2657766 0.6827747 0.4989354 0.6208984 0.8734746 0.6714321 0.3109286
#> [23,] 0.3960574 0.8415093 0.6635124 0.7549983 0.9449139 0.8695222 0.3192816
#> [24,] 0.2671084 0.7671778 0.5309734 0.7063973 0.9263038 0.7601574 0.3131459
#> [25,] 0.3747507 0.8223059 0.6389449 0.7388093 0.9389220 0.8551071 0.3188207
#> [26,] 0.2082594 0.6552315 0.3877832 0.5040966 0.8062110 0.7293136 0.1512935
#> [27,] 0.2521652 0.7026102 0.4995986 0.6049464 0.8588003 0.7033843 0.2995794
#> [28,] 0.2749162 0.7253107 0.4965668 0.6025381 0.8347527 0.7390077 0.1871758
#> [29,] 0.4193505 0.8438753 0.6601558 0.7545292 0.9370673 0.8568341 0.3824287
#> [30,] 0.3749031 0.7866190 0.6393116 0.6906977 0.9004861 0.8108751 0.3542490
#> [31,] 0.3160606 0.8029266 0.5993926 0.7371216 0.9351962 0.8288204 0.2695128
#> [32,] 0.2129112 0.6598536 0.3585868 0.5173783 0.7921096 0.7033924 0.1607466
#> [33,] 0.2817703 0.7458947 0.5235356 0.6585742 0.9050911 0.7233690 0.3278773
#> [34,] 0.3270434 0.7910021 0.5899675 0.6911868 0.9083349 0.8089069 0.2894631
#> [35,] 0.2294218 0.6373673 0.4606153 0.5796165 0.8319154 0.6416600 0.3184746
#> [36,] 0.1860844 0.6869974 0.4338950 0.5738168 0.8515552 0.7515055 0.1602471
#> [37,] 0.2871378 0.7027623 0.5305844 0.6007918 0.8194779 0.7116013 0.3046639
#> [38,] 0.3563088 0.7935775 0.6179526 0.6938284 0.9081477 0.8252175 0.3324161
#> [39,] 0.4107871 0.8259149 0.6674684 0.7847305 0.9485270 0.7998353 0.4312948
#> [40,] 0.3010623 0.7574000 0.5624385 0.7043973 0.9140384 0.7569496 0.3122714
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8243621 0.5332427 0.6534948
#>  [2,] 0.7973578 0.5728436 0.6688094
#>  [3,] 0.6761297 0.3706890 0.5118161
#>  [4,] 0.7154233 0.5209809 0.6209486
#>  [5,] 0.7297030 0.4032052 0.5609454
#>  [6,] 0.8575416 0.5694240 0.7234169
#>  [7,] 0.8909043 0.5922883 0.7535447
#>  [8,] 0.7141254 0.3530331 0.4670619
#>  [9,] 0.8835028 0.6468393 0.7515930
#> [10,] 0.8734835 0.5404397 0.7121059
#> [11,] 0.7726162 0.4666925 0.6141955
#> [12,] 0.8342175 0.5874767 0.6766772
#> [13,] 0.7869843 0.3981041 0.5432893
#> [14,] 0.8848458 0.6419653 0.7359604
#> [15,] 0.8404343 0.6114937 0.7475426
#> [16,] 0.8553351 0.6124589 0.7433688
#> [17,] 0.8386529 0.5933013 0.7101014
#> [18,] 0.8518952 0.5287556 0.6354066
#> [19,] 0.8557636 0.5544509 0.6446153
#> [20,] 0.8067895 0.5058667 0.6130842
#> [21,] 0.8561460 0.5817086 0.7170925
#> [22,] 0.7930489 0.4141656 0.5515340
#> [23,] 0.8778228 0.6580264 0.7689380
#> [24,] 0.8409958 0.4847097 0.6684562
#> [25,] 0.8539491 0.6240170 0.7437527
#> [26,] 0.6998308 0.3433210 0.4691014
#> [27,] 0.7797374 0.4257579 0.6070166
#> [28,] 0.7447413 0.4752903 0.5695283
#> [29,] 0.8809177 0.6514694 0.7449428
#> [30,] 0.8253810 0.5965104 0.7083639
#> [31,] 0.8491583 0.5878633 0.7141540
#> [32,] 0.6792438 0.3525296 0.4506958
#> [33,] 0.8488533 0.4729129 0.6176124
#> [34,] 0.8234127 0.5748194 0.6904091
#> [35,] 0.7377223 0.3914704 0.5385232
#> [36,] 0.7131696 0.4010166 0.5556278
#> [37,] 0.7320889 0.4766049 0.5954007
#> [38,] 0.8411886 0.5985566 0.6994111
#> [39,] 0.8776561 0.6112737 0.7561926
#> [40,] 0.8103628 0.5132297 0.6510709
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6227488 0.3772512
#>  [2,] 0.6177091 0.3822909
#>  [3,] 0.6076857 0.3923143
#>  [4,] 0.6154429 0.3845571
#>  [5,] 0.6116801 0.3883199
#>  [6,] 0.6264423 0.3735577
#>  [7,] 0.6255981 0.3744019
#>  [8,] 0.6119724 0.3880276
#>  [9,] 0.6250427 0.3749573
#> [10,] 0.6270856 0.3729144
#> [11,] 0.6190108 0.3809892
#> [12,] 0.6193045 0.3806955
#> [13,] 0.6174784 0.3825216
#> [14,] 0.6241523 0.3758477
#> [15,] 0.6240214 0.3759786
#> [16,] 0.6260084 0.3739916
#> [17,] 0.6210903 0.3789097
#> [18,] 0.6231258 0.3768742
#> [19,] 0.6223517 0.3776483
#> [20,] 0.6196780 0.3803220
#> [21,] 0.6228872 0.3771128
#> [22,] 0.6162432 0.3837568
#> [23,] 0.6258941 0.3741059
#> [24,] 0.6226444 0.3773556
#> [25,] 0.6246311 0.3753689
#> [26,] 0.6109181 0.3890819
#> [27,] 0.6160086 0.3839914
#> [28,] 0.6155028 0.3844972
#> [29,] 0.6237918 0.3762082
#> [30,] 0.6202840 0.3797160
#> [31,] 0.6250420 0.3749580
#> [32,] 0.6094368 0.3905632
#> [33,] 0.6198071 0.3801929
#> [34,] 0.6216733 0.3783267
#> [35,] 0.6122128 0.3877872
#> [36,] 0.6158979 0.3841021
#> [37,] 0.6126447 0.3873553
#> [38,] 0.6211881 0.3788119
#> [39,] 0.6243907 0.3756093
#> [40,] 0.6211991 0.3788009
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]      [,2]         [,3]       [,4]     [,5]      [,6]
#>  [1,] -1.0943612 1.3359482  0.014869624 0.61234110 2.293629 1.7314332
#>  [2,] -0.6588075 1.1085202  0.369673907 0.76888500 2.087060 1.2591482
#>  [3,] -1.1905456 0.4663208 -0.186693296 0.24328682 1.302997 0.3660466
#>  [4,] -0.9160977 0.9652759  0.104878156 0.32966548 1.703970 1.3175031
#>  [5,] -1.1429131 0.6311241 -0.056290812 0.29093729 1.518463 0.6054334
#>  [6,] -0.9014519 1.6499182  0.177803085 1.01654866 2.796772 1.9094301
#>  [7,] -0.4620163 1.6506531  0.571734015 1.36842132 3.049902 1.5027323
#>  [8,] -1.5351578 0.7164910 -0.608890797 0.04786795 1.497573 1.0951918
#>  [9,] -0.3839988 1.6377241  0.693418016 1.16837605 2.782245 1.7985630
#> [10,] -0.9548576 1.6337678  0.166168913 1.04364351 2.884862 1.9029478
#> [11,] -0.9858606 0.9656002  0.092623065 0.65566604 2.051829 0.9820146
#> [12,] -0.3130803 1.2189633  0.545157378 0.81572692 2.216656 1.3299380
#> [13,] -1.4568506 0.7482270 -0.285786457 0.38297266 1.939828 1.0293524
#> [14,] -0.3428215 1.6193380  0.657929882 1.16439340 2.770474 1.6883281
#> [15,] -0.6110588 1.5336584  0.477900303 0.96269917 2.596939 1.7069662
#> [16,] -0.5993133 1.5045862  0.561134539 1.04513359 2.822778 1.6961253
#> [17,] -0.5223811 1.3701761  0.570984511 0.82869505 2.248890 1.5188614
#> [18,] -0.9757757 1.3711487  0.170820360 0.80014976 2.419381 1.5445541
#> [19,] -0.8041546 1.3066140  0.304215908 0.77166231 2.380694 1.4716934
#> [20,] -0.8524586 1.0987323  0.050659296 0.66204912 2.141084 1.1667419
#> [21,] -0.6331532 1.3759957  0.443290808 1.04060262 2.533651 1.3972192
#> [22,] -1.0161574 0.7665527 -0.004258532 0.49336307 1.932036 0.7146693
#> [23,] -0.4219198 1.6695014  0.678986373 1.12545054 2.842197 1.8967408
#> [24,] -1.0093433 1.1924433  0.124052543 0.87795046 2.531251 1.1535428
#> [25,] -0.5118895 1.5320493  0.570787776 1.03978923 2.732581 1.7752321
#> [26,] -1.3354493 0.6421152 -0.456639902 0.01638690 1.425575 0.9911431
#> [27,] -1.0870976 0.8597584 -0.001605591 0.42611848 1.805361 0.8634661
#> [28,] -0.9698212 0.9709593 -0.013733105 0.41605195 1.619693 1.0408176
#> [29,] -0.3254401 1.6873496  0.663988514 1.12291597 2.700689 1.7892403
#> [30,] -0.5112393 1.3046646  0.572377559 0.80338322 2.202638 1.4557066
#> [31,] -0.7719352 1.4046870  0.402935092 1.03106110 2.669392 1.5772899
#> [32,] -1.3074661 0.6626420 -0.581503130 0.06954102 1.337689 0.8635047
#> [33,] -0.9356971 1.0768354  0.094212076 0.65694689 2.255118 0.9612347
#> [34,] -0.7215879 1.3309765  0.363830983 0.80567341 2.293471 1.4429231
#> [35,] -1.2115792 0.5639556 -0.157865671 0.32119936 1.599263 0.5825761
#> [36,] -1.4756562 0.7861188 -0.265976829 0.29744107 1.746851 1.1066580
#> [37,] -0.9093260 0.8604866  0.122490412 0.40876543 1.512815 0.9031738
#> [38,] -0.5914212 1.3466260  0.480866948 0.81807915 2.291225 1.5521043
#> [39,] -0.3607125 1.5569474  0.696757227 1.29344942 2.913853 1.3852650
#> [40,] -0.8422446 1.1384775  0.251064420 0.86832613 2.363972 1.1360278
#>             [,7]      [,8]        [,9]       [,10]
#>  [1,] -1.6182740 1.5461851  0.13316728  0.63443673
#>  [2,] -0.4288808 1.3698615  0.29346260  0.70280507
#>  [3,] -0.6482403 0.7360418 -0.52926214  0.04727305
#>  [4,] -1.3239118 0.9218715  0.08397300  0.49357641
#>  [5,] -0.7540324 0.9931161 -0.39212790  0.24499975
#>  [6,] -1.2813370 1.7950198  0.27950133  0.96147438
#>  [7,] -0.3911360 2.1000115  0.37343307  1.11760746
#>  [8,] -1.8651377 0.9155054 -0.60573327 -0.13194357
#>  [9,] -0.6076262 2.0260272  0.60517451  1.10712664
#> [10,] -1.3170389 1.9321166  0.16211294  0.90563413
#> [11,] -1.0458576 1.2231430 -0.13342766  0.46498265
#> [12,] -0.4533553 1.6158178  0.35354400  0.73854324
#> [13,] -1.3290228 1.3068425 -0.41337114  0.17359170
#> [14,] -0.4825080 2.0391418  0.58390430  1.02507746
#> [15,] -0.8101472 1.6614628  0.45359536  1.08554862
#> [16,] -0.9843198 1.7770730  0.45765985  1.06355235
#> [17,] -0.6607198 1.6482388  0.37762975  0.89587674
#> [18,] -1.2527346 1.7495434  0.11514950  0.55548227
#> [19,] -1.0498205 1.7805407  0.21867090  0.59545295
#> [20,] -1.0508113 1.4292826  0.02346783  0.46029523
#> [21,] -0.5580084 1.7836421  0.32979134  0.93008474
#> [22,] -0.7957814 1.3434024 -0.34677118  0.20687053
#> [23,] -0.7570752 1.9719725  0.65451149  1.20232407
#> [24,] -0.7854527 1.6656562 -0.06118015  0.70121102
#> [25,] -0.7591966 1.7659166  0.50663374  1.06556572
#> [26,] -1.7244914 0.8464922 -0.64852939 -0.12375219
#> [27,] -0.8493013 1.2641365 -0.29918032  0.43478873
#> [28,] -1.4684668 1.0707592 -0.09891950  0.27992695
#> [29,] -0.4792525 2.0011492  0.62550434  1.07181975
#> [30,] -0.6004138 1.5532389  0.39094606  0.88745112
#> [31,] -0.9970956 1.7280144  0.35513929  0.91564519
#> [32,] -1.6526837 0.7502990 -0.60793839 -0.19785990
#> [33,] -0.7178013 1.7256354 -0.10845438  0.47942627
#> [34,] -0.8979929 1.5396424  0.30154204  0.80203243
#> [35,] -0.7607908 1.0341633 -0.44113554  0.15439895
#> [36,] -1.6563908 0.9108280 -0.40123110  0.22343626
#> [37,] -0.8251863 1.0052468 -0.09364876  0.38633754
#> [38,] -0.6972778 1.6670989  0.39945465  0.84449521
#> [39,] -0.2765704 1.9704193  0.45266925  1.13191771
#> [40,] -0.7895218 1.4523695  0.05293107  0.62374985
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3654403 -0.1357916
#>  [2,] 0.3463349 -0.1335010
#>  [3,] 0.3069259 -0.1306685
#>  [4,] 0.3469318 -0.1233179
#>  [5,] 0.3178208 -0.1365592
#>  [6,] 0.3759231 -0.1410612
#>  [7,] 0.3675543 -0.1458241
#>  [8,] 0.3281594 -0.1274511
#>  [9,] 0.3741427 -0.1368650
#> [10,] 0.3745683 -0.1451658
#> [11,] 0.3470341 -0.1383175
#> [12,] 0.3550478 -0.1315494
#> [13,] 0.3359809 -0.1428783
#> [14,] 0.3710349 -0.1361755
#> [15,] 0.3690608 -0.1375916
#> [16,] 0.3758338 -0.1392968
#> [17,] 0.3591523 -0.1350264
#> [18,] 0.3643573 -0.1384796
#> [19,] 0.3627167 -0.1368251
#> [20,] 0.3538260 -0.1343556
#> [21,] 0.3604090 -0.1414121
#> [22,] 0.3317395 -0.1418931
#> [23,] 0.3780894 -0.1365528
#> [24,] 0.3514098 -0.1493779
#> [25,] 0.3719986 -0.1372532
#> [26,] 0.3244095 -0.1267634
#> [27,] 0.3304393 -0.1422015
#> [28,] 0.3466106 -0.1238925
#> [29,] 0.3710784 -0.1345953
#> [30,] 0.3568079 -0.1339460
#> [31,] 0.3709193 -0.1400857
#> [32,] 0.3246056 -0.1203397
#> [33,] 0.3432998 -0.1454299
#> [34,] 0.3614556 -0.1352010
#> [35,] 0.3181415 -0.1384818
#> [36,] 0.3374530 -0.1347198
#> [37,] 0.3310001 -0.1274428
#> [38,] 0.3599912 -0.1346031
#> [39,] 0.3649475 -0.1432792
#> [40,] 0.3536806 -0.1409602
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3772512 -0.3772512
#>  [2,] -0.6177091  0.6177091
#>  [3,]  0.3923143 -0.3923143
#>  [4,] -0.6154429  0.6154429
#>  [5,]  0.3883199 -0.3883199
#>  [6,] -0.6264423  0.6264423
#>  [7,] -0.6255981  0.6255981
#>  [8,]  0.3880276 -0.3880276
#>  [9,]  0.3749573 -0.3749573
#> [10,] -0.6270856  0.6270856
#> [11,] -0.6190108  0.6190108
#> [12,] -0.6193045  0.6193045
#> [13,] -0.6174784  0.6174784
#> [14,]  0.3758477 -0.3758477
#> [15,] -0.6240214  0.6240214
#> [16,] -0.6260084  0.6260084
#> [17,] -0.6210903  0.6210903
#> [18,]  0.3768742 -0.3768742
#> [19,]  0.3776483 -0.3776483
#> [20,] -0.6196780  0.6196780
#> [21,] -0.6228872  0.6228872
#> [22,] -0.6162432  0.6162432
#> [23,] -0.6258941  0.6258941
#> [24,] -0.6226444  0.6226444
#> [25,] -0.6246311  0.6246311
#> [26,]  0.3890819 -0.3890819
#> [27,] -0.6160086  0.6160086
#> [28,] -0.6155028  0.6155028
#> [29,] -0.6237918  0.6237918
#> [30,] -0.6202840  0.6202840
#> [31,] -0.6250420  0.6250420
#> [32,]  0.3905632 -0.3905632
#> [33,]  0.3801929 -0.3801929
#> [34,]  0.3783267 -0.3783267
#> [35,] -0.6122128  0.6122128
#> [36,] -0.6158979  0.6158979
#> [37,]  0.3873553 -0.3873553
#> [38,]  0.3788119 -0.3788119
#> [39,] -0.6243907  0.6243907
#> [40,]  0.3788009 -0.3788009
#> 
#> $L
#> [1] 0.6935034 0.6937699 0.6924324 0.6930499 0.6904157 0.7870487
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5942029