Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12          V13
#>  [1,] -0.2739314 -0.15522353 -0.10630495 -0.106352132  0.009077480
#>  [2,] -0.2096683 -0.01647250 -0.01212644 -0.081933104 -0.009731759
#>  [3,] -0.3302342 -0.12532240 -0.07294215 -0.112851428 -0.021936085
#>  [4,] -0.2703766 -0.10614538 -0.09543514 -0.223984137 -0.131821788
#>  [5,] -0.2862740 -0.12912974 -0.03954550 -0.004950571 -0.064401974
#>  [6,] -0.3053129 -0.06309922 -0.14233507 -0.140126979 -0.101276308
#>  [7,] -0.1981750 -0.10908624 -0.05059762 -0.058891337 -0.040646580
#>  [8,] -0.3804678 -0.12438254 -0.09169452  0.006512948 -0.153441580
#>  [9,] -0.2709825 -0.05420056 -0.05846518 -0.125758181  0.007227090
#> [10,] -0.3789490 -0.07411428 -0.09222140 -0.076827638  0.041583862
#>                 V14          V15         V16          V17         V18
#>  [1,]  0.0058956792  0.018416045  0.01062829  0.131446613  0.19288553
#>  [2,]  0.0184427957 -0.033668736  0.14711397  0.079748248  0.02036341
#>  [3,] -0.0708648852  0.099040948  0.14428755 -0.032510556  0.04905226
#>  [4,] -0.1433846662 -0.057669077 -0.20397585 -0.166837898 -0.14214406
#>  [5,] -0.0393510660 -0.095132722 -0.07850777 -0.010952917 -0.04092832
#>  [6,]  0.0135504032 -0.007639497  0.04318563  0.066135169  0.10653415
#>  [7,] -0.0401714439 -0.044900070  0.05475739  0.129808245  0.15139954
#>  [8,] -0.0007017639 -0.103434805  0.02510783 -0.017522100  0.05302980
#>  [9,] -0.1230352332 -0.051047900 -0.03692824 -0.000873982 -0.06200314
#> [10,] -0.0420332801  0.145537372  0.07986134  0.233658552  0.09283513
#>               V19         V2         V20         V21          V22        V23
#>  [1,]  0.14210736 -0.3235162  0.10149368  0.17719954  0.076533525 0.05313763
#>  [2,]  0.02124443 -0.2794331  0.11834915  0.01906080  0.165886533 0.14398914
#>  [3,]  0.05657395 -0.2310531  0.10716190  0.10731161  0.240125478 0.16173263
#>  [4,] -0.06675831 -0.2827902 -0.15296433 -0.08615489 -0.111518532 0.11336670
#>  [5,]  0.05252171 -0.3424130 -0.05360297  0.04443700  0.131490751 0.10811638
#>  [6,]  0.11416816 -0.2897293  0.09487882  0.25502192  0.267083030 0.14066669
#>  [7,]  0.16706958 -0.3196983  0.17378173  0.11439864  0.173304055 0.10641720
#>  [8,]  0.06539164 -0.2139231 -0.06487115  0.09016693 -0.006101107 0.03371860
#>  [9,]  0.06123843 -0.1817300  0.10379599  0.03765314  0.058264355 0.17919139
#> [10,]  0.21321088 -0.3223923  0.25247409  0.12123362  0.173418484 0.25450801
#>               V24        V25         V26         V27          V28          V29
#>  [1,]  0.06558318 0.03557694  0.10674337  0.08737250 -0.005703423 -0.052282972
#>  [2,]  0.11149517 0.06896077  0.05699336  0.13512904  0.046675305 -0.002204188
#>  [3,]  0.26302920 0.23191441  0.28731754  0.20076656  0.071632519  0.047019934
#>  [4,] -0.06101664 0.03384764  0.05046692 -0.02416082  0.098880301  0.105943947
#>  [5,]  0.02822541 0.18712848  0.20916120  0.02686180  0.184307908  0.066841840
#>  [6,]  0.31732799 0.25759766  0.22119552  0.21318249  0.153679625  0.161169937
#>  [7,]  0.17997953 0.05417239  0.15204606  0.21211763  0.163464228  0.083485732
#>  [8,]  0.07558584 0.13165702  0.12777848  0.25740211  0.238898443  0.168620988
#>  [9,] -0.01619986 0.12460493 -0.04444203  0.06448087 -0.055715085 -0.016394256
#> [10,]  0.17741541 0.08660276  0.08706710  0.11278264  0.132950665 -0.015392560
#>               V3         V30         V31          V32         V33          V34
#>  [1,] -0.2387863  0.01579650  0.05974958  0.058494921 -0.06203641  0.002169069
#>  [2,] -0.2623932  0.05579233  0.07241857 -0.010456145  0.07414026 -0.037873240
#>  [3,] -0.2378501  0.05314830 -0.01270584 -0.029131950  0.03152736 -0.015877090
#>  [4,] -0.3060718  0.07520385  0.06771920  0.018750503 -0.09088547 -0.092126596
#>  [5,] -0.2290650 -0.04033472  0.05632841 -0.094724779 -0.14308268 -0.150130693
#>  [6,] -0.4048758  0.14936514  0.05345118  0.001943477  0.04656974 -0.057469417
#>  [7,] -0.2402880  0.01333713 -0.05237467  0.023323754 -0.13613634 -0.124542668
#>  [8,] -0.2384433  0.13252393  0.10694490 -0.030065083  0.02458864  0.032208034
#>  [9,] -0.1619293  0.04415186 -0.07115541 -0.021148724 -0.04818755  0.051307793
#> [10,] -0.2459719  0.12076617  0.08654964 -0.030546180 -0.01636941 -0.022375363
#>               V35          V36          V37          V38           V39
#>  [1,] -0.11083837  0.056757827  0.081499927  0.046644217 -0.0917264210
#>  [2,] -0.13161864 -0.174366804 -0.151972221 -0.093492982 -0.1048146435
#>  [3,] -0.05361020 -0.002278640 -0.014443838 -0.020175843 -0.0298741329
#>  [4,] -0.11973569 -0.190760512 -0.024921824 -0.171458593 -0.0861081736
#>  [5,] -0.05101883 -0.104419385 -0.136522570 -0.098570676 -0.1540965178
#>  [6,]  0.04021733  0.011796389 -0.039737896 -0.086266431 -0.0008081841
#>  [7,] -0.01754864 -0.009383556 -0.018076373 -0.005437079 -0.0440667623
#>  [8,] -0.12378689 -0.011422910 -0.080963560 -0.152091664 -0.0912412071
#>  [9,]  0.03290261 -0.017109994 -0.005231555  0.003770605  0.0051419443
#> [10,]  0.02589229 -0.010200657 -0.058499412  0.073211885  0.0342595437
#>               V4           V40          V41         V42         V43         V44
#>  [1,] -0.1965176  0.0209071170 -0.038764354 -0.10809313 -0.09429062 -0.13924653
#>  [2,] -0.3054762 -0.0945714053 -0.018164410 -0.01837356 -0.06310771 -0.08122715
#>  [3,] -0.2375668 -0.0673513235 -0.007537931 -0.12776623 -0.03170319 -0.13025253
#>  [4,] -0.2191688 -0.2143500090 -0.062749569 -0.08094765 -0.15649484 -0.16216091
#>  [5,] -0.3280735 -0.0411373379 -0.171785733 -0.19298652 -0.03488182 -0.10400889
#>  [6,] -0.2484954 -0.0376207701 -0.018665616  0.02995608 -0.03436846 -0.04982399
#>  [7,] -0.2861468 -0.1645273401 -0.141409327 -0.03579348 -0.09063847 -0.12160038
#>  [8,] -0.3617800 -0.0155394919 -0.177216031 -0.02312527 -0.07579167 -0.07031289
#>  [9,] -0.2000810  0.0847148263 -0.101606780 -0.09801131 -0.09912885 -0.11040878
#> [10,] -0.2404761  0.0007848053 -0.005318577 -0.05103600 -0.05069525 -0.08768762
#>               V45         V46         V47         V48        V49         V5
#>  [1,] -0.11122136 -0.04221727 -0.07690797 -0.25985405 -0.2106064 -0.1015576
#>  [2,] -0.09261199 -0.18166025 -0.16844747 -0.26128057 -0.3054910 -0.2164556
#>  [3,] -0.02934157 -0.18202321 -0.15281612 -0.16019603 -0.2549203 -0.2399024
#>  [4,] -0.08949251 -0.12952092 -0.17681779 -0.20614273 -0.2476974 -0.2180979
#>  [5,] -0.17788284 -0.22486357 -0.11445297 -0.18557856 -0.2781826 -0.2265325
#>  [6,] -0.18200795 -0.08757162 -0.21631098 -0.14539389 -0.3599574 -0.2983785
#>  [7,] -0.19983772 -0.07334031 -0.17456929 -0.24194840 -0.2663702 -0.2659772
#>  [8,] -0.03650191 -0.14929698 -0.24600810 -0.28820357 -0.2860927 -0.2457231
#>  [9,] -0.15451800 -0.11396652 -0.06241624 -0.09521741 -0.0997091 -0.1256469
#> [10,] -0.15709629 -0.15443155 -0.18657857 -0.24971134 -0.2695651 -0.2707015
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2984554 -0.3632522 -0.2581855 -0.2679341 -0.3789567 -0.2604782
#>  [2,] -0.3014530 -0.3366794 -0.3895116 -0.2846005 -0.3261749 -0.3935835
#>  [3,] -0.2426144 -0.3529941 -0.3690740 -0.2732997 -0.2771302 -0.3727630
#>  [4,] -0.2654143 -0.3582505 -0.2405910 -0.3767030 -0.3092312 -0.3707455
#>  [5,] -0.3733859 -0.3457258 -0.2142708 -0.2675118 -0.2360912 -0.2911005
#>  [6,] -0.3597746 -0.4383313 -0.4179179 -0.3438453 -0.2910331 -0.3347551
#>  [7,] -0.2399561 -0.2494184 -0.3277360 -0.3581293 -0.3885179 -0.2232387
#>  [8,] -0.3770905 -0.2783201 -0.4404083 -0.3287901 -0.3606204 -0.3251067
#>  [9,] -0.3354801 -0.2721694 -0.2747368 -0.3614854 -0.2075063 -0.2405587
#> [10,] -0.2699580 -0.2740246 -0.3101445 -0.3851133 -0.4223980 -0.3768543
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.2158671 -0.2288570 -0.2480235 -0.3613939 -0.17213988 -0.3177534
#>  [2,] -0.3749714 -0.3205633 -0.4146146 -0.2998477 -0.19811944 -0.3926434
#>  [3,] -0.3612246 -0.3454953 -0.2954382 -0.4220714 -0.28441071 -0.3517285
#>  [4,] -0.3684501 -0.2612862 -0.3089149 -0.3183006 -0.15428703 -0.2281740
#>  [5,] -0.2922873 -0.2867767 -0.3290144 -0.2845393 -0.13350022 -0.3787051
#>  [6,] -0.5019372 -0.5146813 -0.3755372 -0.3711999 -0.25838017 -0.4510619
#>  [7,] -0.3231661 -0.3167000 -0.2305821 -0.3544895 -0.15779861 -0.2576099
#>  [8,] -0.3533835 -0.3641059 -0.4342733 -0.4310976 -0.24455281 -0.2629259
#>  [9,] -0.2139695 -0.3523887 -0.3752911 -0.3376881 -0.25215658 -0.3007811
#> [10,] -0.3406901 -0.3455374 -0.4082522 -0.3334345 -0.08857197 -0.4035212
#>                V7          V8          V9
#>  [1,] -0.16479691 -0.08876088 -0.17976785
#>  [2,] -0.16381253 -0.22227549 -0.06765230
#>  [3,] -0.09154322 -0.17789317 -0.08719183
#>  [4,] -0.11952212 -0.15547231 -0.24894332
#>  [5,] -0.25396062 -0.06323312 -0.17516255
#>  [6,] -0.19623021 -0.14561163 -0.08209365
#>  [7,] -0.07406951 -0.11252947 -0.11943148
#>  [8,] -0.23569676 -0.09241566 -0.10834905
#>  [9,] -0.16474692 -0.20417283 -0.09164630
#> [10,] -0.09738290 -0.23449272 -0.10614010
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]        [,4]       [,5]         [,6]
#> [1,] 0.01221582 -0.04513276  0.06529879 -0.11637643 0.05269008 -0.044902928
#> [2,] 0.04223597  0.07256410 -0.06334545  0.08198235 0.03902167  0.002274084
#>            [,7]        [,8]        [,9]      [,10]
#> [1,] 0.07551725 -0.09686578 -0.07107843 0.01403691
#> [2,] 0.08709934  0.01495823  0.03563451 0.05925206
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  8.402332e-05  0.0005716974  0.0005098074  0.0006249339  0.0007098766
#>  [2,]  1.862625e-04  0.0016081873  0.0020828725  0.0022790719  0.0020494294
#>  [3,] -1.404511e-05 -0.0003255358 -0.0007160990 -0.0005871517 -0.0002960049
#>  [4,]  1.176344e-04  0.0012311393  0.0017788995  0.0019821125  0.0015651075
#>  [5,]  6.082173e-05  0.0003613660  0.0002936054  0.0003651924  0.0004961835
#>  [6,]  8.231967e-05  0.0006246643  0.0006746988  0.0007603559  0.0007304308
#>  [7,]  9.174207e-05  0.0006539517  0.0006583532  0.0007818830  0.0008297255
#>  [8,]  1.680358e-04  0.0013968266  0.0017286340  0.0018889810  0.0017335579
#>  [9,]  1.395500e-04  0.0011900642  0.0014573327  0.0016242118  0.0014852806
#> [10,]  9.310084e-05  0.0006927889  0.0007419374  0.0008551478  0.0008447569
#>                V14          V15           V16           V17           V18
#>  [1,] 0.0008952812 0.0010173210  1.303313e-03  1.424795e-03  1.636366e-03
#>  [2,] 0.0016263038 0.0010358683  1.061094e-03  1.039716e-03  1.132514e-03
#>  [3,] 0.0001855119 0.0007550683  1.104329e-03  1.358848e-03  1.711605e-03
#>  [4,] 0.0008675160 0.0001532144 -1.985731e-05 -6.970852e-05 -7.198664e-05
#>  [5,] 0.0007247489 0.0008986918  1.162251e-03  1.303146e-03  1.488673e-03
#>  [6,] 0.0007386927 0.0006877978  8.317951e-04  8.225231e-04  9.089576e-04
#>  [7,] 0.0009350956 0.0009747252  1.204977e-03  1.292127e-03  1.484090e-03
#>  [8,] 0.0014778371 0.0010599408  1.166699e-03  1.157463e-03  1.275082e-03
#>  [9,] 0.0012680206 0.0009222330  1.033916e-03  1.074339e-03  1.226688e-03
#> [10,] 0.0008732434 0.0008367899  1.011577e-03  1.035313e-03  1.162773e-03
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0018908642  1.162002e-04 0.0022530356 0.0023442638 0.0022109467
#>  [2,] 0.0023398744  2.595014e-04 0.0034982674 0.0038284681 0.0038077823
#>  [3,] 0.0011315394 -2.597748e-05 0.0007059473 0.0005736034 0.0004578931
#>  [4,] 0.0009250511  1.677717e-04 0.0018194002 0.0021176609 0.0022034543
#>  [5,] 0.0015718537  7.998514e-05 0.0017580578 0.0018124710 0.0017356919
#>  [6,] 0.0012926450  1.170916e-04 0.0016951831 0.0018097949 0.0016920686
#>  [7,] 0.0017996716  1.262516e-04 0.0021844670 0.0022923204 0.0021946824
#>  [8,] 0.0022998301  2.351977e-04 0.0033086424 0.0035791461 0.0035036634
#>  [9,] 0.0020827168  1.945940e-04 0.0029728559 0.0032137095 0.0031373684
#> [10,] 0.0015604740  1.315549e-04 0.0019960449 0.0021223999 0.0020126683
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.002575690 0.002711933 0.002454698 0.002587377 0.002678792 0.002731335
#>  [2,] 0.003652330 0.003795814 0.003428256 0.003535796 0.003844501 0.003865890
#>  [3,] 0.001170346 0.001274781 0.001163193 0.001258148 0.001226519 0.001373881
#>  [4,] 0.001849311 0.001871820 0.001616630 0.001605541 0.001932015 0.001954193
#>  [5,] 0.002045482 0.002157596 0.002029672 0.002119338 0.002140218 0.002127752
#>  [6,] 0.001844161 0.001963931 0.001717891 0.001811120 0.001948316 0.002031858
#>  [7,] 0.002492841 0.002637575 0.002397436 0.002505589 0.002611889 0.002673400
#>  [8,] 0.003477020 0.003626589 0.003235750 0.003341043 0.003591181 0.003644592
#>  [9,] 0.003155384 0.003276144 0.002945613 0.003076071 0.003307459 0.003346000
#> [10,] 0.002218543 0.002347159 0.002093737 0.002205729 0.002339599 0.002408687
#>               V29            V3         V30         V31          V32
#>  [1,] 0.002589801  1.275501e-04 0.002427875 0.002331377 2.114549e-03
#>  [2,] 0.003563163  2.562652e-04 0.002999991 0.001849332 1.506513e-03
#>  [3,] 0.001467662 -1.679270e-06 0.001537104 0.002199613 2.095518e-03
#>  [4,] 0.001781700  1.536090e-04 0.001228651 0.000219029 7.447987e-05
#>  [5,] 0.002008997  9.292527e-05 0.001894658 0.001887336 1.703883e-03
#>  [6,] 0.001901689  1.234811e-04 0.001728101 0.001461017 1.296815e-03
#>  [7,] 0.002544783  1.386154e-04 0.002348951 0.002147436 1.927792e-03
#>  [8,] 0.003380764  2.305181e-04 0.002939778 0.002026768 1.687509e-03
#>  [9,] 0.003103213  1.900406e-04 0.002693546 0.001947513 1.660240e-03
#> [10,] 0.002271641  1.413388e-04 0.002068286 0.001795154 1.610542e-03
#>                 V33           V34           V35           V36           V37
#>  [1,]  0.0022879668  0.0024050787  0.0023864672  0.0024642352  0.0023059924
#>  [2,]  0.0014299659  0.0006258513 -0.0001431564  0.0001523557  0.0006656513
#>  [3,]  0.0024617938  0.0032071724  0.0037355691  0.0037354168  0.0031331368
#>  [4,] -0.0002016822 -0.0011907308 -0.0019253481 -0.0015556316 -0.0008301264
#>  [5,]  0.0018444681  0.0020030121  0.0020370776  0.0020330721  0.0018384589
#>  [6,]  0.0013736805  0.0012836670  0.0011140489  0.0012134898  0.0012577363
#>  [7,]  0.0020824421  0.0020999839  0.0019983571  0.0020802974  0.0019924797
#>  [8,]  0.0016970441  0.0011279723  0.0005452255  0.0008007288  0.0011434924
#>  [9,]  0.0016607579  0.0011846897  0.0007130175  0.0009598583  0.0012379572
#> [10,]  0.0017005059  0.0016236216  0.0014608386  0.0015661227  0.0015819043
#>                 V38          V39            V4          V40          V41
#>  [1,]  0.0018915352 0.0015468202  1.486939e-04 1.614340e-03 0.0012882929
#>  [2,]  0.0010974422 0.0013232919  3.475951e-04 1.014324e-03 0.0011218175
#>  [3,]  0.0022069381 0.0015537358 -3.806955e-05 1.903575e-03 0.0013009040
#>  [4,] -0.0001548576 0.0002760387  2.279627e-04 8.395996e-06 0.0002656230
#>  [5,]  0.0014373011 0.0011155293  1.044165e-04 1.217058e-03 0.0009666183
#>  [6,]  0.0011640306 0.0010727762  1.544536e-04 1.052661e-03 0.0009270641
#>  [7,]  0.0016963452 0.0014481498  1.657429e-04 1.490867e-03 0.0012430672
#>  [8,]  0.0013722712 0.0014467021  3.117290e-04 1.216107e-03 0.0012176498
#>  [9,]  0.0013659944 0.0013710326  2.575060e-04 1.187207e-03 0.0011090880
#> [10,]  0.0014178800 0.0012739767  1.698057e-04 1.259195e-03 0.0010814329
#>                V42           V43           V44           V45           V46
#>  [1,] 0.0010222325  0.0007642162  0.0005804177  0.0004196401  0.0003416893
#>  [2,] 0.0018844484  0.0020020453  0.0016635095  0.0019778762  0.0014470968
#>  [3,] 0.0003003709 -0.0002518289 -0.0003032398 -0.0008275074 -0.0005222611
#>  [4,] 0.0011797641  0.0014602247  0.0012614255  0.0016078529  0.0011356875
#>  [5,] 0.0006777118  0.0004607843  0.0003517779  0.0001834002  0.0001762312
#>  [6,] 0.0009472349  0.0008157986  0.0006258531  0.0006494311  0.0005234433
#>  [7,] 0.0010978086  0.0008713578  0.0006680342  0.0005652232  0.0004642821
#>  [8,] 0.0017371377  0.0017488823  0.0014226182  0.0016422143  0.0012178158
#>  [9,] 0.0014870108  0.0014842838  0.0012276200  0.0013540915  0.0009729785
#> [10,] 0.0010757204  0.0009161631  0.0007028805  0.0006842844  0.0005429019
#>                 V47           V48           V49            V5          V50
#>  [1,]  0.0002555532  0.0002045777  1.315728e-04  1.873347e-04 8.124853e-05
#>  [2,]  0.0008916727  0.0008256202  5.025426e-04  4.708151e-04 1.373146e-04
#>  [3,] -0.0002065385 -0.0002762332 -1.588654e-04 -5.443492e-05 2.129478e-05
#>  [4,]  0.0006991391  0.0006813326  4.183110e-04  3.116705e-04 7.458146e-05
#>  [5,]  0.0001574713  0.0001047257  6.611411e-05  1.342925e-04 5.799378e-05
#>  [6,]  0.0003435950  0.0002901755  1.797481e-04  1.968027e-04 7.119072e-05
#>  [7,]  0.0003321164  0.0002660854  1.648676e-04  2.165903e-04 8.279817e-05
#>  [8,]  0.0007558619  0.0006989331  4.287876e-04  4.138402e-04 1.315518e-04
#>  [9,]  0.0006074989  0.0005712024  3.529998e-04  3.411413e-04 1.110873e-04
#> [10,]  0.0003652598  0.0003085466  1.903051e-04  2.171160e-04 7.960306e-05
#>                 V51           V52          V53          V54          V55
#>  [1,]  5.241790e-05  3.748123e-05 3.793482e-05 4.676343e-05 3.975128e-05
#>  [2,]  1.374724e-04  1.187372e-04 5.854368e-05 7.550463e-05 6.411429e-05
#>  [3,] -2.230131e-05 -3.663515e-05 1.469984e-05 1.390738e-05 1.128966e-05
#>  [4,]  9.688754e-05  8.443886e-05 2.277346e-05 4.255822e-05 2.853424e-05
#>  [5,]  3.262368e-05  2.271724e-05 2.818693e-05 3.258042e-05 2.816428e-05
#>  [6,]  5.716216e-05  4.584165e-05 3.129275e-05 3.941158e-05 3.402583e-05
#>  [7,]  5.910730e-05  4.454871e-05 3.876983e-05 4.673525e-05 3.986038e-05
#>  [8,]  1.222285e-04  1.042687e-04 5.725236e-05 7.347652e-05 6.259255e-05
#>  [9,]  1.014404e-04  8.425437e-05 4.792884e-05 6.362585e-05 5.270677e-05
#> [10,]  6.201272e-05  4.943094e-05 3.613676e-05 4.488271e-05 3.841005e-05
#>                V56          V57           V58          V59           V6
#>  [1,] 2.717883e-05 3.756396e-05  2.747410e-05 2.736808e-05 0.0003955737
#>  [2,] 3.959815e-05 4.568722e-05  6.154682e-05 5.349634e-05 0.0006317461
#>  [3,] 1.035650e-05 2.096661e-05 -6.899570e-06 1.286119e-06 0.0001363329
#>  [4,] 1.009889e-05 1.305613e-05  3.578231e-05 3.086544e-05 0.0004170128
#>  [5,] 2.123125e-05 2.921466e-05  1.786616e-05 1.867053e-05 0.0002859787
#>  [6,] 2.289651e-05 2.936801e-05  2.848005e-05 2.592218e-05 0.0003254172
#>  [7,] 2.795148e-05 3.662518e-05  2.943714e-05 2.890624e-05 0.0003974323
#>  [8,] 3.925073e-05 4.702316e-05  5.727622e-05 4.978259e-05 0.0005949947
#>  [9,] 3.157740e-05 3.959487e-05  4.637088e-05 4.140533e-05 0.0005325834
#> [10,] 2.618902e-05 3.340300e-05  3.111907e-05 2.922849e-05 0.0003741798
#>                 V60           V7            V8            V9
#>  [1,]  3.081703e-05 0.0005377717  4.689594e-04  0.0004394397
#>  [2,]  2.089345e-05 0.0007321498  9.705348e-04  0.0013107838
#>  [3,]  3.127431e-05 0.0002734881 -2.486549e-05 -0.0003385845
#>  [4,] -3.065764e-06 0.0004154066  6.640752e-04  0.0009744992
#>  [5,]  2.445758e-05 0.0003953503  3.159600e-04  0.0002879134
#>  [6,]  2.061814e-05 0.0004190311  4.625236e-04  0.0005040509
#>  [7,]  2.846557e-05 0.0005227848  4.995705e-04  0.0005095985
#>  [8,]  2.511141e-05 0.0007196584  8.879018e-04  0.0011405607
#>  [9,]  2.273369e-05 0.0006441425  7.552523e-04  0.0009585009
#> [10,]  2.452855e-05 0.0004799826  5.083767e-04  0.0005459355
#> 
#> $vW[[2]]
#>              [,1]         [,2]         [,3]         [,4]        [,5]
#> [1,]  0.004813439 -0.008075497 -0.006804061  0.004765982 -0.00351209
#> [2,] -0.004813439  0.008075497  0.006804061 -0.004765982  0.00351209
#>              [,6]         [,7]         [,8]         [,9]       [,10]
#> [1,] -0.009142993 -0.006491444 -0.004096426  0.005629529 -0.00552839
#> [2,]  0.009142993  0.006491444  0.004096426 -0.005629529  0.00552839
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.28541191  0.29475090  0.37622933 -0.01823739  0.20786695  0.63033735
#>  [7]  0.35571408  0.45658731  0.15027063  0.54109669
#> 
#> $B[[2]]
#> [1] 0.003287007 0.095799784
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.004067483 0.004133004 0.003271041 0.001112942 0.003292979 0.002748204
#>  [7] 0.003859857 0.004248204 0.003908410 0.003313257
#> 
#> $vB[[2]]
#> [1] -0.008969037  0.008969037
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [2,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [3,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#>  [4,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#>  [5,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>  [6,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#>  [7,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [8,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#>  [9,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [10,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [11,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [12,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [13,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [14,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [15,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [16,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [17,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [18,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [19,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [20,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [21,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [22,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [23,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [24,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [25,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [26,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [27,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [28,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [29,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [30,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [31,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [32,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [33,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [34,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [35,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [36,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [37,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [38,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [39,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [40,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [2,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [3,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#>  [4,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#>  [5,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>  [6,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#>  [7,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [8,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#>  [9,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [10,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [11,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [12,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [13,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [14,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [15,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [16,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [17,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [18,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [19,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [20,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [21,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [22,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [23,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [24,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [25,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [26,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [27,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [28,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [29,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [30,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [31,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [32,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [33,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [34,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [35,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [36,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [37,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [38,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [39,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [40,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [2,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [3,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#>  [4,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#>  [5,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>  [6,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#>  [7,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [8,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#>  [9,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [10,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [11,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [12,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [13,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [14,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [15,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [16,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [17,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [18,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [19,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [20,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [21,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [22,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [23,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [24,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [25,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [26,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [27,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [28,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [29,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [30,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [31,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [32,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [33,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [34,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [35,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [36,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [37,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [38,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [39,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [40,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [2,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [3,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#>  [4,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#>  [5,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>  [6,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#>  [7,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [8,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#>  [9,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [10,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [11,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [12,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [13,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [14,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [15,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [16,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [17,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [18,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [19,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [20,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [21,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [22,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [23,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [24,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [25,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [26,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [27,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [28,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [29,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [30,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [31,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [32,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [33,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [34,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [35,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [36,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [37,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [38,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [39,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [40,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [2,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [3,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#>  [4,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#>  [5,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>  [6,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#>  [7,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [8,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#>  [9,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [10,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [11,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [12,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [13,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [14,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [15,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [16,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [17,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [18,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [19,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [20,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [21,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [22,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [23,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [24,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [25,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [26,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [27,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [28,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [29,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [30,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [31,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [32,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [33,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [34,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [35,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [36,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [37,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [38,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [39,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [40,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [2,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [3,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#>  [4,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#>  [5,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#>  [6,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#>  [7,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [8,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#>  [9,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [10,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [11,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [12,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [13,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [14,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [15,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [16,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [17,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [18,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [19,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [20,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [21,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [22,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [23,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [24,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [25,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [26,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [27,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [28,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [29,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [30,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [31,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [32,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [33,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [34,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [35,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [36,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [37,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [38,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [39,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [40,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7695932 0.7515292 0.8795082 0.2579132 0.6124171 0.9324582 0.8459511
#>  [2,] 0.6143588 0.6369251 0.7747796 0.3979230 0.5906591 0.8497049 0.6992505
#>  [3,] 0.5938282 0.5794971 0.7661286 0.3076565 0.5024464 0.8573445 0.6655124
#>  [4,] 0.6978960 0.5999287 0.7848868 0.1453332 0.4257924 0.8762427 0.7262049
#>  [5,] 0.7160696 0.6744131 0.8137933 0.2038144 0.5302583 0.8969173 0.7962866
#>  [6,] 0.7126745 0.6754695 0.8312271 0.2158959 0.5608324 0.9019205 0.7984386
#>  [7,] 0.5843856 0.4426870 0.5913904 0.1514331 0.2768520 0.7387046 0.5491099
#>  [8,] 0.7627150 0.7420057 0.8738441 0.2553063 0.6017887 0.9296449 0.8385326
#>  [9,] 0.6899523 0.6609541 0.7936882 0.3151193 0.5781908 0.8757828 0.7665626
#> [10,] 0.7533379 0.7335552 0.8612774 0.2690456 0.6074916 0.9219408 0.8215406
#> [11,] 0.5821586 0.4662388 0.6746182 0.2507517 0.3757612 0.7781735 0.5705700
#> [12,] 0.6762728 0.5947741 0.7751840 0.1538340 0.4265350 0.8593943 0.7321324
#> [13,] 0.6431157 0.5776696 0.7481592 0.2061120 0.4371600 0.8342758 0.6633771
#> [14,] 0.7078034 0.7063530 0.8518415 0.2803924 0.5894754 0.9202637 0.8034965
#> [15,] 0.4995387 0.3929361 0.5841050 0.1229705 0.2459416 0.7211293 0.5083582
#> [16,] 0.6058414 0.6025033 0.7715863 0.1979379 0.4593238 0.8689933 0.6944735
#> [17,] 0.7104527 0.6709540 0.7811637 0.1911477 0.4891763 0.8729515 0.7609512
#> [18,] 0.5739275 0.4957794 0.6420899 0.2839085 0.3728892 0.7727258 0.5541580
#> [19,] 0.6004873 0.6223209 0.7659944 0.2221731 0.5037660 0.8652716 0.7047046
#> [20,] 0.5758657 0.5562721 0.6627438 0.2556835 0.4232830 0.7599095 0.6083687
#> [21,] 0.6520322 0.6542974 0.8107016 0.2329220 0.5327659 0.8808569 0.7407962
#> [22,] 0.6558080 0.5666316 0.8039551 0.2352137 0.4913509 0.8878433 0.7346770
#> [23,] 0.6730103 0.6236285 0.7169521 0.1872055 0.4106347 0.7967749 0.6930276
#> [24,] 0.6949472 0.6567220 0.8108238 0.3204433 0.5814800 0.8877579 0.7729317
#> [25,] 0.5963928 0.5110064 0.7249026 0.2137305 0.4310117 0.8306376 0.6704494
#> [26,] 0.6473157 0.6788543 0.8133414 0.3703446 0.6181413 0.8902918 0.7611198
#> [27,] 0.7060845 0.6757917 0.7751983 0.2463176 0.4959722 0.8733700 0.7580421
#> [28,] 0.5910284 0.5398160 0.6440876 0.3241354 0.4429407 0.7679348 0.6118524
#> [29,] 0.5982429 0.5981036 0.7539156 0.2043228 0.4400652 0.8533876 0.6783184
#> [30,] 0.6732882 0.6929819 0.8365426 0.3283871 0.6242022 0.9100900 0.7905177
#> [31,] 0.6730307 0.6729833 0.8221338 0.3470326 0.5920816 0.9018512 0.7670804
#> [32,] 0.5494721 0.5701408 0.7314066 0.1828708 0.4222719 0.8257519 0.6268285
#> [33,] 0.6608319 0.6807589 0.8165320 0.2982660 0.5678367 0.8874271 0.7439869
#> [34,] 0.7048783 0.7132018 0.8449431 0.2602147 0.5937748 0.9143059 0.8072859
#> [35,] 0.6314800 0.5978535 0.7398296 0.2459593 0.4635073 0.8566722 0.6998148
#> [36,] 0.5564833 0.5313415 0.7119201 0.2546083 0.4455385 0.8028225 0.6236010
#> [37,] 0.5922663 0.5698320 0.7144061 0.1565561 0.3544906 0.8260437 0.6411251
#> [38,] 0.5258860 0.5086818 0.6526819 0.2014649 0.3581215 0.7930126 0.6009405
#> [39,] 0.6975435 0.6620241 0.7965736 0.2535637 0.5447490 0.8803348 0.7745394
#> [40,] 0.6986186 0.6873010 0.8409045 0.2387635 0.5645903 0.9147036 0.7950631
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7636469 0.5757407 0.9116290
#>  [2,] 0.7292240 0.5321380 0.7652741
#>  [3,] 0.7041496 0.5215564 0.7837303
#>  [4,] 0.6205316 0.5222778 0.8540602
#>  [5,] 0.6816562 0.5230609 0.8824569
#>  [6,] 0.6908122 0.5321074 0.8786478
#>  [7,] 0.5278373 0.4385591 0.7499100
#>  [8,] 0.7557555 0.5738918 0.9075387
#>  [9,] 0.7432009 0.5272719 0.8309583
#> [10,] 0.7432064 0.5718233 0.8956234
#> [11,] 0.6197753 0.4660935 0.7073557
#> [12,] 0.6260795 0.4719056 0.8540408
#> [13,] 0.5863009 0.4915055 0.8068756
#> [14,] 0.7628601 0.5384889 0.8796063
#> [15,] 0.4752334 0.3860753 0.6875945
#> [16,] 0.6910223 0.4586853 0.8097050
#> [17,] 0.6669770 0.5090186 0.8656532
#> [18,] 0.6032281 0.5005420 0.7199237
#> [19,] 0.6936330 0.4860167 0.8116580
#> [20,] 0.6244405 0.4441493 0.7277839
#> [21,] 0.7069788 0.5050216 0.8300375
#> [22,] 0.6976831 0.5386063 0.8236827
#> [23,] 0.6067983 0.4568442 0.8348160
#> [24,] 0.7369702 0.5695398 0.8419295
#> [25,] 0.6439830 0.5080845 0.7712522
#> [26,] 0.7606080 0.5309730 0.8199159
#> [27,] 0.6656557 0.5394998 0.8662291
#> [28,] 0.6542229 0.5078225 0.7340584
#> [29,] 0.6779883 0.4542212 0.7963976
#> [30,] 0.7709452 0.5490638 0.8476939
#> [31,] 0.7729182 0.5392032 0.8406098
#> [32,] 0.5773208 0.4993509 0.7805904
#> [33,] 0.7249362 0.5152122 0.8369950
#> [34,] 0.7460423 0.5292201 0.8812610
#> [35,] 0.7095458 0.4918375 0.8030434
#> [36,] 0.5945535 0.4509838 0.7325181
#> [37,] 0.5783554 0.4299543 0.7958944
#> [38,] 0.5946520 0.3883242 0.7289041
#> [39,] 0.6787033 0.5421753 0.8640184
#> [40,] 0.7203667 0.5365220 0.8783190
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.3893795 0.6106205
#>  [2,] 0.3906703 0.6093297
#>  [3,] 0.3970870 0.6029130
#>  [4,] 0.4044653 0.5955347
#>  [5,] 0.3973926 0.6026074
#>  [6,] 0.3968484 0.6031516
#>  [7,] 0.4134789 0.5865211
#>  [8,] 0.3900705 0.6099295
#>  [9,] 0.3917897 0.6082103
#> [10,] 0.3902170 0.6097830
#> [11,] 0.4071860 0.5928140
#> [12,] 0.4053123 0.5946877
#> [13,] 0.4050584 0.5949416
#> [14,] 0.3910500 0.6089500
#> [15,] 0.4207000 0.5793000
#> [16,] 0.4024562 0.5975438
#> [17,] 0.3986931 0.6013069
#> [18,] 0.4035565 0.5964435
#> [19,] 0.3999319 0.6000681
#> [20,] 0.4039913 0.5960087
#> [21,] 0.3979220 0.6020780
#> [22,] 0.3998722 0.6001278
#> [23,] 0.4036243 0.5963757
#> [24,] 0.3909817 0.6090183
#> [25,] 0.4048860 0.5951140
#> [26,] 0.3889184 0.6110816
#> [27,] 0.3952564 0.6047436
#> [28,] 0.3983066 0.6016934
#> [29,] 0.4028254 0.5971746
#> [30,] 0.3893136 0.6106864
#> [31,] 0.3891968 0.6108032
#> [32,] 0.4071190 0.5928810
#> [33,] 0.3933350 0.6066650
#> [34,] 0.3923116 0.6076884
#> [35,] 0.3983913 0.6016087
#> [36,] 0.4062144 0.5937856
#> [37,] 0.4088097 0.5911903
#> [38,] 0.4095154 0.5904846
#> [39,] 0.3954652 0.6045348
#> [40,] 0.3946530 0.6053470
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]        [,2]      [,3]       [,4]         [,5]      [,6]
#>  [1,]  1.206015851  1.10678480 1.9877814 -1.0568429  0.457483998 2.6250772
#>  [2,]  0.465671908  0.56204274 1.2354989 -0.4141269  0.366690642 1.7322883
#>  [3,]  0.379813760  0.32070928 1.1865787 -0.8110981  0.009785863 1.7934073
#>  [4,]  0.837298558  0.40516804 1.2943752 -1.7716824 -0.299039198 1.9573211
#>  [5,]  0.925048446  0.72821347 1.4748490 -1.3626222  0.121181396 2.1634325
#>  [6,]  0.908409252  0.73302870 1.5943492 -1.2897456  0.244540912 2.2187483
#>  [7,]  0.340803025 -0.23026417 0.3697162 -1.7234050 -0.960130928 1.0392467
#>  [8,]  1.167622518  1.05641919 1.9353834 -1.0705088  0.412923571 2.5812480
#>  [9,]  0.799896186  0.66754875 1.3473022 -0.7762934  0.315350886 1.9530869
#> [10,]  1.116494405  1.01273582 1.8259407 -0.9994700  0.436780208 2.4690131
#> [11,]  0.331640930 -0.13525075 0.7291477 -1.0946070 -0.507579172 1.2550540
#> [12,]  0.736695495  0.38373688 1.2378180 -1.7048413 -0.296002359 1.8102680
#> [13,]  0.588912834  0.31321416 1.0888186 -1.3485230 -0.252696178 1.6162393
#> [14,]  0.884739521  0.87773695 1.7491180 -0.9425161  0.361797201 2.4459358
#> [15,] -0.001845329 -0.43498718 0.3396480 -1.9645961 -1.120375920 0.9500702
#> [16,]  0.429864878  0.41590647 1.2172900 -1.3992328 -0.163065200 1.8920869
#> [17,]  0.897583860  0.71250311 1.2724607 -1.4425702 -0.043301383 1.9273109
#> [18,]  0.297893716 -0.01688268 0.5844465 -0.9251562 -0.519841916 1.2237669
#> [19,]  0.407495804  0.49941112 1.1858299 -1.2530471  0.015064314 1.8597828
#> [20,]  0.305824317  0.22604595 0.6755458 -1.0685261 -0.309310839 1.1521832
#> [21,]  0.627984039  0.63798310 1.4545753 -1.1918847  0.131251566 2.0005697
#> [22,]  0.644668589  0.26812098 1.4111998 -1.1791020 -0.034599773 2.0688978
#> [23,]  0.721831952  0.50497823 0.9293928 -1.4682712 -0.361342064 1.3662582
#> [24,]  0.823350807  0.64872024 1.4553720 -0.7517352  0.328852013 2.0680408
#> [25,]  0.390457516  0.04403265 0.9689120 -1.3025839 -0.277724649 1.5901531
#> [26,]  0.607260579  0.74851147 1.4718699 -0.5307387  0.481666604 2.0937248
#> [27,]  0.876442815  0.73449872 1.2378999 -1.1183494 -0.016111629 1.9310896
#> [28,]  0.368218463  0.15960196 0.5931500 -0.7348314 -0.229235732 1.1966866
#> [29,]  0.398149170  0.39756971 1.1196058 -1.3594926 -0.240897262 1.7614215
#> [30,]  0.723095227  0.81409720 1.6327251 -0.7154888  0.507423326 2.3147345
#> [31,]  0.721924620  0.72170913 1.5308716 -0.6321083  0.372577426 2.2179648
#> [32,]  0.198538075  0.28242559 1.0017706 -1.4970175 -0.313453884 1.5558139
#> [33,]  0.667003829  0.75726150 1.4930257 -0.8555687  0.273030462 2.0647254
#> [34,]  0.870637224  0.91098545 1.6954769 -1.0448527  0.379592251 2.3673811
#> [35,]  0.538571130  0.39652915 1.0450833 -1.1202803 -0.146230839 1.7879207
#> [36,]  0.226901737  0.12553076 0.9047280 -1.0741835 -0.218713905 1.4040291
#> [37,]  0.373342298  0.28116556 0.9168805 -1.6840788 -0.599357910 1.5578435
#> [38,]  0.103636689  0.03473053 0.6308490 -1.3771639 -0.583526687 1.3431811
#> [39,]  0.835627193  0.67232718 1.3650154 -1.0796952  0.179476100 1.9956047
#> [40,]  0.840728221  0.78753135 1.6649736 -1.1594708  0.259812750 2.3724679
#>            [,7]        [,8]         [,9]     [,10]
#>  [1,] 1.7031913  1.17277884  0.305312623 2.3336894
#>  [2,] 0.8437315  0.99068912  0.128729542 1.1818157
#>  [3,] 0.6879576  0.86713670  0.086279074 1.2875389
#>  [4,] 0.9754524  0.49180539  0.089170170 1.7668073
#>  [5,] 1.3632448  0.76139340  0.092309101 2.0159050
#>  [6,] 1.3765641  0.80391914  0.128606548 1.9796869
#>  [7,] 0.1970750  0.11146456 -0.247011764 1.0981321
#>  [8,] 1.6473505  1.12954800  0.297747409 2.2839456
#>  [9,] 1.1890024  1.06267253  0.109196107 1.5924345
#> [10,] 1.5268205  1.06270101  0.289293989 2.1495146
#> [11,] 0.2841773  0.48859473 -0.135834557 0.8825757
#> [12,] 1.0054687  0.51543418 -0.112496187 1.7666520
#> [13,] 0.6783802  0.34869425 -0.033981433 1.4298347
#> [14,] 1.4082928  1.16842432  0.154260580 1.9887077
#> [15,] 0.0334358 -0.09914744 -0.463839985 0.7888971
#> [16,] 0.8211173  0.80490297 -0.165636456 1.4480947
#> [17,] 1.1579016  0.69454407  0.036078383 1.8630598
#> [18,] 0.2174852  0.41893390  0.002167963 0.9440829
#> [19,] 0.8698024  0.81715926 -0.055947598 1.4608199
#> [20,] 0.4404604  0.50843926 -0.224338835 0.9834080
#> [21,] 1.0501109  0.88075583  0.020087158 1.5858931
#> [22,] 1.0184831  0.83628898  0.154733070 1.5415004
#> [23,] 0.8143121  0.43387353 -0.173053836 1.6201514
#> [24,] 1.2249400  1.03027988  0.279973814 1.6726551
#> [25,] 0.7102182  0.59269400  0.032340917 1.2153952
#> [26,] 1.1588284  1.15601554  0.124050649 1.5157779
#> [27,] 1.1419751  0.68860146  0.158329077 1.8680206
#> [28,] 0.4551053  0.63765389  0.031292404 1.0153117
#> [29,] 0.7460546  0.74454214 -0.183629608 1.3639295
#> [30,] 1.3280489  1.21365617  0.196888822 1.7166273
#> [31,] 1.1918983  1.22486295  0.157135305 1.6627726
#> [32,] 0.5186348  0.31178446 -0.002596212 1.2691103
#> [33,] 1.0667950  0.96908065  0.060867412 1.6360376
#> [34,] 1.4324699  1.07761482  0.117013768 2.0044256
#> [35,] 0.8464163  0.89317901 -0.032652718 1.4054253
#> [36,] 0.5048611  0.38282166 -0.196696405 1.0074363
#> [37,] 0.5802507  0.31602577 -0.282037714 1.3608292
#> [38,] 0.4093855  0.38323040 -0.454361885 0.9890695
#> [39,] 1.2341234  0.74781910  0.169103134 1.8490741
#> [40,] 1.3557191  0.94628128  0.146348752 1.9766069
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.07881335 0.3711079
#>  [2,] -0.09875137 0.3457443
#>  [3,] -0.09017536 0.3274421
#>  [4,] -0.06188433 0.3250095
#>  [5,] -0.06749770 0.3488434
#>  [6,] -0.06810754 0.3505068
#>  [7,] -0.06561369 0.2839884
#>  [8,] -0.07846387 0.3685522
#>  [9,] -0.08630489 0.3534906
#> [10,] -0.07961752 0.3667825
#> [11,] -0.08035548 0.2952549
#> [12,] -0.05940186 0.3239767
#> [13,] -0.06700576 0.3174263
#> [14,] -0.08274366 0.3601571
#> [15,] -0.05682820 0.2630722
#> [16,] -0.07371989 0.3215216
#> [17,] -0.06881021 0.3421035
#> [18,] -0.08912753 0.3015406
#> [19,] -0.07684038 0.3289085
#> [20,] -0.07942595 0.3094361
#> [21,] -0.07560838 0.3385227
#> [22,] -0.07568024 0.3303175
#> [23,] -0.06545159 0.3249350
#> [24,] -0.08767630 0.3555114
#> [25,] -0.07329453 0.3118531
#> [26,] -0.09428630 0.3575744
#> [27,] -0.07739437 0.3478757
#> [28,] -0.09337077 0.3191552
#> [29,] -0.07508536 0.3186210
#> [30,] -0.08928976 0.3609085
#> [31,] -0.09352016 0.3571696
#> [32,] -0.06921057 0.3066778
#> [33,] -0.08464943 0.3486658
#> [34,] -0.07819571 0.3594100
#> [35,] -0.08310946 0.3290630
#> [36,] -0.07299589 0.3066412
#> [37,] -0.06431743 0.3045709
#> [38,] -0.07111767 0.2948515
#> [39,] -0.07471583 0.3496807
#> [40,] -0.07497730 0.3528177
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.3893795  0.3893795
#>  [2,] -0.3906703  0.3906703
#>  [3,] -0.3970870  0.3970870
#>  [4,] -0.4044653  0.4044653
#>  [5,]  0.6026074 -0.6026074
#>  [6,]  0.6031516 -0.6031516
#>  [7,] -0.4134789  0.4134789
#>  [8,] -0.3900705  0.3900705
#>  [9,] -0.3917897  0.3917897
#> [10,] -0.3902170  0.3902170
#> [11,] -0.4071860  0.4071860
#> [12,] -0.4053123  0.4053123
#> [13,] -0.4050584  0.4050584
#> [14,]  0.6089500 -0.6089500
#> [15,]  0.5793000 -0.5793000
#> [16,]  0.5975438 -0.5975438
#> [17,] -0.3986931  0.3986931
#> [18,] -0.4035565  0.4035565
#> [19,]  0.6000681 -0.6000681
#> [20,]  0.5960087 -0.5960087
#> [21,]  0.6020780 -0.6020780
#> [22,]  0.6001278 -0.6001278
#> [23,] -0.4036243  0.4036243
#> [24,] -0.3909817  0.3909817
#> [25,] -0.4048860  0.4048860
#> [26,]  0.6110816 -0.6110816
#> [27,]  0.6047436 -0.6047436
#> [28,] -0.3983066  0.3983066
#> [29,]  0.5971746 -0.5971746
#> [30,]  0.6106864 -0.6106864
#> [31,] -0.3891968  0.3891968
#> [32,]  0.5928810 -0.5928810
#> [33,] -0.3933350  0.3933350
#> [34,]  0.6076884 -0.6076884
#> [35,]  0.6016087 -0.6016087
#> [36,]  0.5937856 -0.5937856
#> [37,]  0.5911903 -0.5911903
#> [38,]  0.5904846 -0.5904846
#> [39,] -0.3954652  0.3954652
#> [40,]  0.6053470 -0.6053470
#> 
#> $L
#> [1] 0.6891171 0.7193611 0.7383414 0.6840088 0.8274458 0.7119164
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174