Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.3308395 -0.07188392 -0.19507529 -0.05481033 -0.023411824  0.01588454
#>  [2,] -0.3104655 -0.05084229 -0.07716412 -0.20705994 -0.081402255 -0.05834827
#>  [3,] -0.2933985 -0.17883101 -0.12226790  0.02867597 -0.002123266  0.04446000
#>  [4,] -0.2918885 -0.09329102 -0.03436318 -0.12097031 -0.070782579 -0.10388627
#>  [5,] -0.2602982 -0.18124704 -0.06165635 -0.14481798 -0.153154994 -0.10229502
#>  [6,] -0.2087030 -0.07336550 -0.05296971 -0.06665554 -0.141478894 -0.08397484
#>  [7,] -0.3139649 -0.16594909 -0.13411634 -0.05948041 -0.131381939 -0.09243120
#>  [8,] -0.3829638 -0.09497196 -0.07391435 -0.06900690 -0.097068041 -0.01344017
#>  [9,] -0.2143978 -0.10656110 -0.15457619 -0.14612902 -0.123989853 -0.06279836
#> [10,] -0.3005354 -0.17765860 -0.04756250 -0.08459730 -0.045079549 -0.13849277
#>                V15          V16           V17          V18          V19
#>  [1,] -0.062963578 -0.029456826  0.0312042687  0.067345294  0.111163336
#>  [2,] -0.103463508 -0.109412564 -0.1124437774  0.064213401  0.122353585
#>  [3,] -0.082503615  0.015644031  0.0245588726  0.118677844  0.001646570
#>  [4,] -0.055421888  0.131479246  0.0715592102  0.130742990  0.020985539
#>  [5,] -0.092485749 -0.012306930 -0.0006901034  0.038457926 -0.005672833
#>  [6,] -0.084098454 -0.127855215  0.0369418920 -0.107396497  0.020034586
#>  [7,] -0.071444927 -0.102486625 -0.0925466044 -0.006958922  0.044852679
#>  [8,] -0.071214060 -0.009516438  0.0838153422  0.034897762  0.001978980
#>  [9,]  0.027749163  0.121319444  0.1208742875  0.030288995  0.148560640
#> [10,]  0.008134869  0.021248141  0.0394577521  0.014741490  0.092374283
#>               V2         V20         V21        V22          V23        V24
#>  [1,] -0.2531947  0.19985282  0.20892321 0.17440379  0.198729674 0.15731589
#>  [2,] -0.3913876  0.02838936  0.19850465 0.08792459  0.165543351 0.19577583
#>  [3,] -0.2491289  0.15726839  0.15674766 0.08458718  0.181272211 0.13323410
#>  [4,] -0.2879447  0.13312499  0.22551491 0.11675045  0.233547767 0.07570825
#>  [5,] -0.2094947  0.10314559  0.11787106 0.12432820  0.077766340 0.10536220
#>  [6,] -0.3045284 -0.04159362 -0.07694951 0.08012822  0.144134761 0.09111033
#>  [7,] -0.3437699 -0.02645913  0.07478850 0.07833367 -0.002160228 0.10798505
#>  [8,] -0.2555160  0.01888939  0.06201065 0.18644619  0.080791128 0.12427872
#>  [9,] -0.2397658  0.12019591  0.13424478 0.17994106  0.044019817 0.03991988
#> [10,] -0.2946721  0.20004418  0.21866920 0.24574978  0.119927913 0.11064010
#>               V25        V26          V27         V28         V29         V3
#>  [1,]  0.09074413 0.24273604  0.276003797  0.29119075  0.09526203 -0.4024773
#>  [2,]  0.20254920 0.20900108  0.306075291  0.16558059  0.14002375 -0.2790336
#>  [3,]  0.08912191 0.19992308  0.138832707  0.16528044  0.14764358 -0.3272934
#>  [4,]  0.10286624 0.07104395  0.120461142  0.06089563  0.17427959 -0.2112951
#>  [5,]  0.07258636 0.01788516 -0.069008238 -0.14506341  0.05342777 -0.2170965
#>  [6,] -0.01364361 0.02325055  0.002408199  0.03576907  0.15053028 -0.2730950
#>  [7,]  0.16651886 0.19114501  0.096916586  0.12769113  0.02531456 -0.3496202
#>  [8,]  0.05956677 0.12633661  0.020157765  0.12753043 -0.03784502 -0.2341462
#>  [9,]  0.02591592 0.06520242  0.054841335  0.07273048  0.01694029 -0.2309521
#> [10,]  0.15707822 0.20727905  0.089134798  0.23497185  0.10309930 -0.2530052
#>                V30         V31         V32          V33         V34
#>  [1,]  0.130556309  0.13310945  0.06075041  0.041364866  0.01597274
#>  [2,]  0.170591534  0.09040044  0.11592908  0.028578830  0.09883251
#>  [3,]  0.080451747  0.05331866  0.10556404 -0.021300064 -0.07700144
#>  [4,]  0.038121685 -0.03014561 -0.01910362 -0.071290250 -0.11389338
#>  [5,] -0.019621215 -0.16098052 -0.06469117 -0.102653221 -0.02848986
#>  [6,]  0.041489456  0.06655840 -0.03827585  0.058643239  0.11224286
#>  [7,] -0.011280781  0.11058100  0.09218636  0.040530668  0.06871531
#>  [8,] -0.008293816 -0.06907852 -0.05481794 -0.095491218 -0.02356708
#>  [9,]  0.052937672  0.07517400 -0.03465152 -0.034500285 -0.04145561
#> [10,]  0.098203098 -0.04181679 -0.09999746 -0.002018618 -0.06376870
#>                V35          V36         V37          V38         V39         V4
#>  [1,]  0.177123024  0.130123885 -0.02093637  0.019736201 -0.02927637 -0.2842391
#>  [2,] -0.052110051  0.036387632  0.04270818 -0.071788053  0.02314417 -0.3858695
#>  [3,] -0.130556600 -0.157419264 -0.06514870 -0.128061602 -0.09430424 -0.3166493
#>  [4,]  0.012475944 -0.060857509 -0.14316319 -0.028721188 -0.04718768 -0.1848427
#>  [5,] -0.141496145 -0.108372946 -0.07051945 -0.217648895 -0.20102384 -0.2850662
#>  [6,] -0.075440565  0.008515224 -0.05877015 -0.092629211 -0.15121184 -0.1701499
#>  [7,]  0.004451158 -0.092313414 -0.07222509  0.039608179 -0.02331395 -0.2591302
#>  [8,] -0.194339298 -0.032286207  0.02039723 -0.133629518 -0.05471413 -0.2383083
#>  [9,]  0.053985130  0.088859917 -0.10640164  0.005202688 -0.01198681 -0.2925578
#> [10,] -0.113089703 -0.076956952 -0.08312881 -0.081660430 -0.15890080 -0.3350923
#>               V40          V41         V42         V43         V44         V45
#>  [1,] -0.09640528 -0.035500882  0.01931182  0.01446409 -0.10942592 -0.09756074
#>  [2,]  0.05049421 -0.003946563 -0.05725947 -0.05529231 -0.07219208 -0.10936344
#>  [3,] -0.14324757 -0.119177373 -0.06243230 -0.07073854 -0.08910601 -0.15548855
#>  [4,] -0.07369064 -0.024236994 -0.15188584 -0.02293756 -0.12058100 -0.07761858
#>  [5,] -0.19211995 -0.151361533 -0.19985312 -0.12886223 -0.08717776 -0.22414990
#>  [6,]  0.03635642 -0.109879993 -0.14447467 -0.19322771 -0.16377285 -0.19043217
#>  [7,] -0.14162300 -0.136852177 -0.16898493 -0.07974580 -0.18830700 -0.19904626
#>  [8,] -0.09330397 -0.038778993 -0.07066099 -0.22398648 -0.16334321 -0.19524186
#>  [9,]  0.02799303  0.036941447 -0.04516804 -0.06853571 -0.13718398 -0.06750983
#> [10,] -0.07940274 -0.130738267 -0.04343888 -0.07093970 -0.05262753 -0.15377899
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.14596738 -0.1915384 -0.2857446 -0.2868255 -0.2656015 -0.3783063
#>  [2,] -0.18528130 -0.1470512 -0.2827846 -0.3516521 -0.1992944 -0.3553780
#>  [3,] -0.09932059 -0.1119501 -0.1642840 -0.2667355 -0.2496734 -0.3909146
#>  [4,] -0.20175374 -0.2589823 -0.1500858 -0.3129055 -0.2814558 -0.2369409
#>  [5,] -0.16423369 -0.1279628 -0.3186322 -0.2163968 -0.3087805 -0.2699901
#>  [6,] -0.06144016 -0.2344742 -0.2151396 -0.2918320 -0.2876210 -0.3190585
#>  [7,] -0.15127511 -0.2197636 -0.1612119 -0.3008517 -0.1925480 -0.3119867
#>  [8,] -0.27263036 -0.2782238 -0.2509855 -0.3139206 -0.3294131 -0.2453510
#>  [9,] -0.16567082 -0.2243830 -0.1720954 -0.2779932 -0.1442497 -0.2849883
#> [10,] -0.24959943 -0.1532901 -0.2847903 -0.2116257 -0.2371752 -0.3607591
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.4611847 -0.4273562 -0.3666309 -0.4576685 -0.3656269 -0.4580337
#>  [2,] -0.3513330 -0.3060778 -0.3897138 -0.2964346 -0.3141160 -0.3867650
#>  [3,] -0.3930517 -0.3684508 -0.3039496 -0.3890042 -0.4459139 -0.4084591
#>  [4,] -0.3998858 -0.3204579 -0.3336795 -0.2113760 -0.3418919 -0.2252898
#>  [5,] -0.3177644 -0.2103982 -0.3543363 -0.3513597 -0.3194300 -0.3015252
#>  [6,] -0.3088223 -0.2614877 -0.4268261 -0.3467861 -0.4098779 -0.3678566
#>  [7,] -0.2943900 -0.3116737 -0.2252536 -0.3767455 -0.2367228 -0.3365829
#>  [8,] -0.2280620 -0.2347331 -0.4019357 -0.2501659 -0.3068689 -0.2613028
#>  [9,] -0.2661017 -0.3227638 -0.1810946 -0.1794611 -0.3087218 -0.2383628
#> [10,] -0.3169442 -0.3826073 -0.3791258 -0.3491816 -0.3415288 -0.2638483
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.4804207 -0.4977304 -0.4077746 -0.1306213 -0.4173063 -0.2490409
#>  [2,] -0.3195908 -0.3968365 -0.4786596 -0.2859604 -0.3564616 -0.1609475
#>  [3,] -0.3072379 -0.2641820 -0.3192118 -0.2748587 -0.3411932 -0.1621915
#>  [4,] -0.2319651 -0.2394575 -0.2788243 -0.2614139 -0.4086824 -0.2415470
#>  [5,] -0.2136929 -0.2355543 -0.3323115 -0.1867357 -0.3166176 -0.1747747
#>  [6,] -0.3244667 -0.3652334 -0.2312487 -0.2768449 -0.2786253 -0.2393674
#>  [7,] -0.3260014 -0.2825095 -0.3712781 -0.2444410 -0.3504780 -0.1382030
#>  [8,] -0.4003762 -0.4026112 -0.2772625 -0.1797802 -0.3676435 -0.2028920
#>  [9,] -0.3571199 -0.2172371 -0.2075941 -0.1076777 -0.1862888 -0.1179017
#> [10,] -0.3963003 -0.3834837 -0.4085683 -0.1415144 -0.2711743 -0.1428838
#>               V8          V9
#>  [1,] -0.1286448 -0.17169268
#>  [2,] -0.2322096 -0.23521626
#>  [3,] -0.2146308 -0.04585082
#>  [4,] -0.2096439 -0.01607272
#>  [5,] -0.1835934 -0.14752852
#>  [6,] -0.2075977 -0.23258968
#>  [7,] -0.2021487 -0.18040982
#>  [8,] -0.1329484 -0.05868288
#>  [9,] -0.1550624 -0.06738739
#> [10,] -0.1597257 -0.15067639
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]       [,5]        [,6]
#> [1,]  0.05748241 -0.08314445 -0.03339001  0.02924175 0.06081077 -0.19286983
#> [2,] -0.01972005  0.04419464  0.04546660 -0.05902919 0.07545476  0.02143423
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.12747308 -0.08509207 -0.13711130 -0.08376513
#> [2,]  0.06015634  0.04904957  0.01686924  0.03344307
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.536557e-04 -5.585912e-04 -6.335487e-04 -3.864427e-04 -3.377992e-04
#>  [2,] -1.511101e-05 -5.943812e-05 -5.577905e-05 -2.041370e-05 -1.504471e-05
#>  [3,] -7.148765e-06 -9.447391e-06  1.483804e-05  6.055578e-05  7.238008e-05
#>  [4,] -1.780164e-04 -7.259125e-04 -8.042294e-04 -4.791776e-04 -3.825261e-04
#>  [5,] -2.479430e-05 -1.132503e-04 -1.359049e-04 -1.194656e-04 -1.014747e-04
#>  [6,]  5.562514e-05  2.096317e-04  2.366224e-04  1.338636e-04  1.075681e-04
#>  [7,]  4.058567e-05  1.495237e-04  1.722376e-04  1.077992e-04  1.052796e-04
#>  [8,]  3.328534e-05  1.369920e-04  1.671853e-04  1.259524e-04  1.145327e-04
#>  [9,]  4.789856e-05  2.004614e-04  2.378772e-04  1.730040e-04  1.530879e-04
#> [10,]  2.561613e-05  1.171958e-04  1.558561e-04  1.401715e-04  1.409812e-04
#>                 V14           V15           V16           V17           V18
#>  [1,] -3.908046e-04 -3.609645e-04 -2.220861e-04  7.158223e-05  1.815242e-04
#>  [2,] -3.233939e-05 -3.647630e-05 -1.911000e-06  5.002943e-05  6.200756e-05
#>  [3,]  4.798828e-05  5.173649e-05  1.178814e-04  2.081364e-04  2.301300e-04
#>  [4,] -2.835538e-04 -8.563505e-05  2.243433e-04  6.694376e-04  8.557466e-04
#>  [5,] -4.709834e-05  4.213599e-06  6.156993e-05  1.165150e-04  1.154957e-04
#>  [6,]  4.054214e-05 -7.879323e-05 -2.255738e-04 -4.265822e-04 -4.608189e-04
#>  [7,]  9.214175e-05  4.752883e-05 -1.313569e-05 -1.226665e-04 -1.432341e-04
#>  [8,]  6.616390e-05  1.057622e-05 -3.655327e-05 -1.058272e-04 -1.159565e-04
#>  [9,]  1.101456e-04  5.233515e-05 -2.189830e-06 -9.327086e-05 -1.234427e-04
#> [10,]  1.086373e-04  8.112346e-05  9.128576e-05  9.974144e-05  9.043947e-05
#>                 V19            V2           V20           V21           V22
#>  [1,] -5.256298e-04 -2.120136e-04 -0.0011477752 -8.876063e-04 -6.244200e-04
#>  [2,] -5.352002e-05 -2.189376e-05 -0.0001426457 -9.489351e-05 -3.800523e-05
#>  [3,]  5.369822e-05 -1.165787e-05 -0.0000618455  6.097146e-05  1.628562e-04
#>  [4,] -2.929685e-04 -2.688097e-04 -0.0010807322 -6.305243e-04 -2.638150e-04
#>  [5,] -1.198735e-04 -4.578001e-05 -0.0001654266 -8.445126e-05 -3.545212e-05
#>  [6,]  1.425766e-04  9.332177e-05  0.0003897213  1.248392e-04  8.854603e-07
#>  [7,]  1.582173e-04  6.009453e-05  0.0003347494  2.127345e-04  1.508017e-04
#>  [8,]  1.199024e-04  5.254644e-05  0.0002219908  1.432335e-04  1.100558e-04
#>  [9,]  1.492841e-04  7.130837e-05  0.0003151288  2.287289e-04  1.805256e-04
#> [10,]  1.074053e-04  3.653561e-05  0.0001397330  1.826968e-04  2.176721e-04
#>                 V23           V24           V25           V26           V27
#>  [1,]  8.638111e-05  0.0007560455  0.0010305558  0.0009938356  0.0008809087
#>  [2,]  8.854687e-05  0.0002136160  0.0002425048  0.0002288943  0.0002200536
#>  [3,]  3.778537e-04  0.0006069910  0.0006446370  0.0006230717  0.0005995372
#>  [4,]  1.010304e-03  0.0022538585  0.0027430754  0.0027196134  0.0025320140
#>  [5,]  2.437888e-04  0.0005461158  0.0007103823  0.0007457241  0.0007076380
#>  [6,] -6.216655e-04 -0.0013024237 -0.0016126754 -0.0016341773 -0.0015319461
#>  [7,] -1.185289e-04 -0.0004155525 -0.0005538400 -0.0005407641 -0.0004913864
#>  [8,] -1.177535e-04 -0.0003606648 -0.0004983160 -0.0005102963 -0.0004721879
#>  [9,] -8.618412e-05 -0.0003527116 -0.0004930957 -0.0004950341 -0.0004459535
#> [10,]  2.241025e-04  0.0002481314  0.0002017183  0.0001834468  0.0001891181
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0010872482  0.0013439811 -1.699184e-04  9.971214e-04  9.940121e-04
#>  [2,]  0.0002280645  0.0002549234 -1.246713e-05  1.888742e-04  1.957849e-04
#>  [3,]  0.0005966644  0.0006339887  6.315020e-06  5.241727e-04  4.782052e-04
#>  [4,]  0.0028294913  0.0032132237 -1.901743e-04  2.640922e-03  2.497646e-03
#>  [5,]  0.0007929232  0.0008343501 -3.172467e-05  7.124463e-04  6.839420e-04
#>  [6,] -0.0017331188 -0.0018952812  5.719882e-05 -1.639411e-03 -1.502278e-03
#>  [7,] -0.0005810819 -0.0006705005  4.349241e-05 -5.699792e-04 -5.298530e-04
#>  [8,] -0.0005706986 -0.0006418147  4.056331e-05 -5.572674e-04 -5.218511e-04
#>  [9,] -0.0005511898 -0.0006514807  5.645211e-05 -5.458429e-04 -5.111860e-04
#> [10,]  0.0001343234  0.0001068714  4.154333e-05  9.543645e-05  6.848429e-05
#>                 V32           V33           V34           V35           V36
#>  [1,]  4.965477e-04  7.479372e-04  1.290134e-03  1.510584e-03  1.943039e-03
#>  [2,]  1.144193e-04  1.341721e-04  2.033487e-04  2.418403e-04  3.032095e-04
#>  [3,]  3.205848e-04  3.565447e-04  4.817166e-04  5.565421e-04  6.247051e-04
#>  [4,]  1.565831e-03  2.134135e-03  3.278132e-03  3.739352e-03  4.220029e-03
#>  [5,]  4.803866e-04  5.815056e-04  8.600905e-04  9.758292e-04  1.025719e-03
#>  [6,] -1.005447e-03 -1.273289e-03 -1.906892e-03 -2.187912e-03 -2.361600e-03
#>  [7,] -3.062262e-04 -4.316342e-04 -7.146884e-04 -8.478720e-04 -9.686345e-04
#>  [8,] -3.376785e-04 -4.609121e-04 -7.261997e-04 -8.352758e-04 -9.077620e-04
#>  [9,] -3.089309e-04 -4.640821e-04 -7.620006e-04 -8.748882e-04 -9.789119e-04
#> [10,]  6.917082e-05  1.521262e-05 -4.879691e-05 -5.360074e-05 -7.790597e-05
#>                 V37           V38           V39            V4           V40
#>  [1,]  1.611799e-03  0.0008226975  7.003009e-04 -2.642688e-04  1.034954e-03
#>  [2,]  2.488302e-04  0.0001150581  8.492800e-05 -2.559509e-05  1.456688e-04
#>  [3,]  5.187188e-04  0.0003087784  2.363625e-04 -7.240796e-06  3.005351e-04
#>  [4,]  3.335635e-03  0.0018419260  1.366496e-03 -3.154991e-04  1.802103e-03
#>  [5,]  7.675091e-04  0.0004078310  2.544572e-04 -5.308616e-05  3.183284e-04
#>  [6,] -1.870162e-03 -0.0011067965 -7.798044e-04  1.032658e-04 -8.967344e-04
#>  [7,] -7.821700e-04 -0.0004379067 -3.121190e-04  6.832648e-05 -3.836176e-04
#>  [8,] -6.992486e-04 -0.0003939441 -2.683317e-04  6.113173e-05 -3.187868e-04
#>  [9,] -7.611419e-04 -0.0004173255 -2.973273e-04  8.429457e-05 -3.778558e-04
#> [10,] -4.431336e-05 -0.0000042726 -7.044558e-07  4.974116e-05 -1.288232e-05
#>                 V41           V42           V43           V44           V45
#>  [1,]  5.613238e-04  1.698101e-04  1.926763e-04 -1.817232e-04 -7.927396e-04
#>  [2,]  8.725691e-05  5.039730e-05  4.701160e-05 -1.029352e-05 -1.022050e-04
#>  [3,]  2.202049e-04  1.905288e-04  1.805901e-04  6.622419e-05 -7.993538e-05
#>  [4,]  9.508240e-04  3.736087e-04  5.045778e-04 -1.590828e-04 -1.271943e-03
#>  [5,]  1.690037e-04  6.553660e-05  1.224024e-04  6.955090e-06 -2.039648e-04
#>  [6,] -5.203435e-04 -2.755858e-04 -3.682373e-04 -2.767738e-05  4.736036e-04
#>  [7,] -1.989395e-04 -6.065326e-05 -1.077446e-04  4.335943e-05  2.806315e-04
#>  [8,] -1.584086e-04 -3.232743e-05 -8.358683e-05  3.301500e-05  2.350087e-04
#>  [9,] -1.755795e-04 -1.857342e-05 -7.138563e-05  7.164172e-05  3.194874e-04
#> [10,]  5.269906e-05  1.246417e-04  9.622297e-05  9.365403e-05  1.349664e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -6.629624e-04 -3.891691e-04 -2.139036e-04 -8.627209e-05 -2.098592e-04
#>  [2,] -8.753851e-05 -4.727656e-05 -2.246427e-05 -9.679411e-06 -2.163626e-05
#>  [3,] -7.182999e-05 -3.037385e-05  2.084051e-06  3.422789e-06  4.523103e-07
#>  [4,] -1.085469e-03 -6.605979e-04 -3.485139e-04 -1.661044e-04 -2.497884e-04
#>  [5,] -1.835687e-04 -1.436644e-04 -8.017938e-05 -4.560274e-05 -5.736801e-05
#>  [6,]  4.119765e-04  2.849642e-04  1.334963e-04  6.776531e-05  9.476040e-05
#>  [7,]  2.389992e-04  1.533000e-04  7.481617e-05  3.510255e-05  6.090768e-05
#>  [8,]  2.018508e-04  1.427146e-04  7.826405e-05  4.060320e-05  5.664784e-05
#>  [9,]  2.718950e-04  1.795522e-04  1.002927e-04  4.910937e-05  7.332095e-05
#> [10,]  1.117165e-04  7.774144e-05  5.794827e-05  2.931828e-05  4.377113e-05
#>                 V50           V51           V52           V53           V54
#>  [1,] -1.016911e-05 -5.742627e-05 -8.602706e-05 -1.787803e-05 -2.750253e-05
#>  [2,] -8.680066e-07 -7.354545e-06 -1.121504e-05 -1.890745e-06 -2.296975e-06
#>  [3,]  6.135430e-06 -3.031334e-06 -1.156011e-05  6.656752e-07  6.654950e-07
#>  [4,] -1.030320e-05 -7.496835e-05 -1.149394e-04 -1.392403e-05 -2.939725e-05
#>  [5,] -3.295284e-06 -1.272635e-05 -1.705826e-05 -7.691170e-07 -4.108589e-06
#>  [6,] -2.873739e-06  2.344646e-05  4.385196e-05  1.878990e-06  9.291195e-06
#>  [7,]  1.916102e-06  1.588763e-05  2.737949e-05  4.447515e-06  8.157403e-06
#>  [8,]  2.914509e-06  1.302689e-05  2.004764e-05  2.456487e-06  6.265813e-06
#>  [9,]  5.071543e-06  1.919174e-05  2.769014e-05  4.489604e-06  9.161444e-06
#> [10,]  7.654687e-06  9.775437e-06  8.759584e-06  3.225890e-06  6.056997e-06
#>                 V55           V56           V57           V58           V59
#>  [1,] -1.295308e-05  2.502429e-06  6.720499e-06 -9.182272e-06 -9.173616e-06
#>  [2,] -5.617061e-07  1.160883e-06  2.118761e-06 -4.676942e-07  1.342456e-07
#>  [3,]  3.720957e-06  5.460871e-06  6.811237e-06  3.625184e-06  5.312801e-06
#>  [4,] -7.337519e-06  1.332357e-05  3.141315e-05 -7.518328e-07  1.905792e-06
#>  [5,]  7.416426e-07  2.272803e-06  9.924350e-06  5.983310e-08  7.429758e-07
#>  [6,] -4.140415e-06 -9.454791e-06 -2.172620e-05 -4.935743e-06 -6.973126e-06
#>  [7,]  1.369318e-06 -2.665617e-06 -7.352430e-06  1.496268e-08 -7.398634e-07
#>  [8,]  1.163366e-06 -1.761862e-06 -7.477609e-06  3.344945e-07  5.373162e-08
#>  [9,]  3.073618e-06 -1.401809e-06 -6.599106e-06  1.529128e-06  1.411968e-06
#> [10,]  5.640312e-06  3.412758e-06  1.860311e-06  4.066304e-06  5.441563e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -5.834116e-05  3.849725e-07  4.298054e-05 -1.389542e-04 -7.137496e-04
#>  [2,]  3.497223e-07  4.186063e-07  1.149953e-05 -1.433085e-05 -8.686155e-05
#>  [3,]  3.918225e-05  4.481941e-06  6.118495e-05  2.792948e-05 -6.381563e-05
#>  [4,]  2.736029e-05  1.427401e-05  1.804659e-04 -1.253765e-04 -9.668129e-04
#>  [5,]  1.258603e-05  3.249939e-06  3.875900e-05 -2.720951e-06 -1.380106e-04
#>  [6,] -5.013519e-05 -1.196181e-05 -1.346851e-04 -2.404982e-05  3.252519e-04
#>  [7,]  2.225371e-06 -3.953109e-06 -4.135541e-05  2.144532e-05  2.085410e-04
#>  [8,] -1.338971e-06 -3.257628e-06 -3.309955e-05  1.729090e-05  1.680333e-04
#>  [9,]  1.202546e-05 -2.709159e-06 -2.598254e-05  4.389874e-05  2.438974e-04
#> [10,]  3.584688e-05  2.340472e-06  3.082149e-05  5.112027e-05  1.054393e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,]  0.07770275  0.07778229  0.0604658  0.05913584  0.03017194  0.05742385
#> [2,] -0.07770275 -0.07778229 -0.0604658 -0.05913584 -0.03017194 -0.05742385
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.06955144  0.04961612  0.05691042  0.06136022
#> [2,] -0.06955144 -0.04961612 -0.05691042 -0.06136022
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.840779269 0.613865453 0.579747630 0.489098841 0.002677316 0.240126249
#>  [7] 0.605900987 0.278844290 0.186934750 0.589716825
#> 
#> $B[[2]]
#> [1] -0.09779871  0.01257600
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0012759395  0.0002750421  0.0007812884  0.0036021892  0.0009268505
#>  [6] -0.0021817899 -0.0007262053 -0.0006954379 -0.0006902156  0.0002002048
#> 
#> $vB[[2]]
#> [1]  0.08270512 -0.08270512
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#>  [2,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#>  [3,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#>  [4,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#>  [5,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#>  [6,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [7,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [8,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#>  [9,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [10,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [11,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [12,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [13,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [14,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [15,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [16,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [17,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [18,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [19,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [20,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [21,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [22,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [23,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [24,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [25,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [26,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [27,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [28,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [29,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [30,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [31,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [32,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [33,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [34,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#> [35,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [36,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [37,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [38,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [39,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [40,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#>  [2,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#>  [3,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#>  [4,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#>  [5,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#>  [6,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [7,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [8,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#>  [9,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [10,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [11,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [12,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [13,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [14,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [15,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [16,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [17,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [18,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [19,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [20,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [21,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [22,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [23,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [24,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [25,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [26,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [27,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [28,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [29,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [30,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [31,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [32,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [33,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [34,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#> [35,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [36,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [37,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [38,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [39,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [40,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#>  [2,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#>  [3,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#>  [4,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#>  [5,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#>  [6,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [7,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [8,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#>  [9,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [10,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [11,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [12,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [13,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [14,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [15,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [16,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [17,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [18,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [19,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [20,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [21,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [22,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [23,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [24,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [25,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [26,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [27,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [28,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [29,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [30,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [31,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [32,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [33,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [34,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#> [35,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [36,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [37,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [38,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [39,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [40,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#>  [2,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#>  [3,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#>  [4,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#>  [5,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#>  [6,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [7,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [8,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#>  [9,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [10,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [11,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [12,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [13,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [14,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [15,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [16,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [17,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [18,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [19,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [20,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [21,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [22,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [23,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [24,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [25,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [26,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [27,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [28,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [29,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [30,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [31,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [32,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [33,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [34,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#> [35,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [36,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [37,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [38,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [39,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [40,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#>  [2,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#>  [3,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#>  [4,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#>  [5,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#>  [6,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [7,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [8,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#>  [9,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [10,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [11,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [12,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [13,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [14,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [15,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [16,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [17,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [18,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [19,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [20,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [21,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [22,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [23,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [24,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [25,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [26,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [27,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [28,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [29,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [30,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [31,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [32,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [33,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [34,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#> [35,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [36,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [37,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [38,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [39,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [40,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#>  [2,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#>  [3,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#>  [4,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#>  [5,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#>  [6,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [7,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [8,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#>  [9,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [10,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [11,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [12,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [13,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [14,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [15,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [16,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [17,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [18,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [19,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [20,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [21,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [22,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [23,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [24,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [25,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [26,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [27,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [28,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [29,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [30,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [31,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [32,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [33,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [34,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#> [35,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [36,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [37,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [38,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [39,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [40,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9081065 0.8276458 0.8048807 0.7867742 0.4005602 0.4726778 0.6323891
#>  [2,] 0.8982297 0.8338477 0.6911284 0.6761138 0.2581897 0.4503970 0.6315049
#>  [3,] 0.8310546 0.7329869 0.6510615 0.5847341 0.2154294 0.3965144 0.5334329
#>  [4,] 0.9113454 0.8471801 0.7435218 0.6901780 0.3050579 0.4572935 0.6622394
#>  [5,] 0.8960238 0.8367799 0.6976258 0.6696125 0.2948648 0.5108147 0.6770889
#>  [6,] 0.8076887 0.6551989 0.4915728 0.5098690 0.1321289 0.2582345 0.4243020
#>  [7,] 0.9191274 0.8718553 0.7680384 0.7345557 0.3383969 0.5541339 0.7080464
#>  [8,] 0.8630753 0.7845702 0.7278607 0.7101567 0.3729621 0.4637986 0.6254815
#>  [9,] 0.8631909 0.7656063 0.7511684 0.7302969 0.3745773 0.4631466 0.6328925
#> [10,] 0.8670975 0.7956529 0.6176067 0.5491428 0.1469644 0.3890185 0.5664499
#> [11,] 0.7924253 0.6370610 0.6402258 0.6488176 0.3146286 0.3743394 0.5170365
#> [12,] 0.9252792 0.8673441 0.8256724 0.7972304 0.3780195 0.4716089 0.6728920
#> [13,] 0.8795189 0.7954391 0.6479120 0.6440147 0.2558485 0.4193771 0.6102901
#> [14,] 0.7755327 0.6359529 0.5478156 0.4762389 0.1463250 0.2936235 0.4382621
#> [15,] 0.7207676 0.6284364 0.4205295 0.4022701 0.1086982 0.2931255 0.4055686
#> [16,] 0.8935230 0.8373596 0.7177530 0.6934827 0.3128113 0.4885490 0.6855364
#> [17,] 0.8946333 0.8460509 0.7541142 0.7038017 0.3127082 0.4763363 0.6653507
#> [18,] 0.9388418 0.9003572 0.8123396 0.7767469 0.3841249 0.5763039 0.7582745
#> [19,] 0.7962899 0.6430775 0.4739012 0.5076914 0.1474201 0.2948391 0.4387892
#> [20,] 0.8892120 0.8393971 0.7734411 0.6980844 0.3179678 0.4994954 0.6821833
#> [21,] 0.9204027 0.8817861 0.7761256 0.7339257 0.3469160 0.5778560 0.7274601
#> [22,] 0.8831174 0.7964681 0.7346101 0.7318144 0.3859104 0.4496694 0.6255703
#> [23,] 0.8207898 0.7433777 0.5259142 0.5211898 0.1465560 0.3134596 0.4701190
#> [24,] 0.8662373 0.8161578 0.7452054 0.6925566 0.3781252 0.5154455 0.6966715
#> [25,] 0.8994243 0.8524095 0.7663662 0.7084819 0.3561441 0.5657821 0.7218422
#> [26,] 0.8923386 0.8226890 0.7413555 0.6774007 0.2172382 0.3741444 0.5821550
#> [27,] 0.8833286 0.8440343 0.7570921 0.7208994 0.3889288 0.5319642 0.6971143
#> [28,] 0.9019866 0.8245137 0.7903318 0.7663522 0.3693235 0.4210805 0.6315570
#> [29,] 0.8928219 0.8462742 0.7913921 0.7258060 0.3424931 0.5049457 0.6974428
#> [30,] 0.9230332 0.8646966 0.7653005 0.7228549 0.2874172 0.5014595 0.7056518
#> [31,] 0.9276058 0.8638117 0.7170463 0.6875963 0.2109212 0.4614058 0.6656007
#> [32,] 0.8928699 0.8105101 0.7461696 0.7387147 0.3540198 0.4290456 0.6072816
#> [33,] 0.8350377 0.7468685 0.5366479 0.5241771 0.1742571 0.4327527 0.5563460
#> [34,] 0.8909667 0.8422387 0.7537348 0.6823059 0.2798876 0.4886649 0.6612652
#> [35,] 0.8554443 0.7482864 0.7148315 0.6799553 0.2729657 0.4276310 0.5620849
#> [36,] 0.8864797 0.7972461 0.7531366 0.7508969 0.3987495 0.4595607 0.6255463
#> [37,] 0.9163576 0.8610373 0.7564599 0.7235719 0.3161169 0.4904827 0.6807515
#> [38,] 0.8747705 0.7783985 0.7039795 0.6959404 0.2931004 0.4074781 0.5578318
#> [39,] 0.9031514 0.8561106 0.7636379 0.6977533 0.3155091 0.5444845 0.7029973
#> [40,] 0.8651366 0.7611275 0.6513971 0.5706719 0.1711946 0.3661715 0.5147495
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5844586 0.6956866 0.8244644
#>  [2,] 0.4621113 0.6059111 0.7150358
#>  [3,] 0.3634448 0.5294968 0.6097295
#>  [4,] 0.4960330 0.6111301 0.7586118
#>  [5,] 0.4843006 0.6044257 0.7157708
#>  [6,] 0.2800322 0.5122534 0.5146634
#>  [7,] 0.5368524 0.6616086 0.7711019
#>  [8,] 0.5331107 0.6122190 0.7498766
#>  [9,] 0.5613504 0.6172640 0.7654782
#> [10,] 0.3350365 0.5091973 0.5696850
#> [11,] 0.4555740 0.5975397 0.6439228
#> [12,] 0.6019646 0.6905132 0.8405452
#> [13,] 0.4439134 0.5872752 0.6855665
#> [14,] 0.2512200 0.4479200 0.5003682
#> [15,] 0.2429521 0.4394887 0.3955689
#> [16,] 0.5399745 0.6406287 0.7246843
#> [17,] 0.5114399 0.6099893 0.7505263
#> [18,] 0.6053257 0.6805040 0.8237765
#> [19,] 0.2928709 0.5199384 0.5163573
#> [20,] 0.5284718 0.5919927 0.7572544
#> [21,] 0.5438379 0.6312535 0.7770738
#> [22,] 0.5540612 0.6615569 0.7616248
#> [23,] 0.2913930 0.5367879 0.5345742
#> [24,] 0.5593443 0.5962504 0.7455840
#> [25,] 0.5453047 0.6165418 0.7559087
#> [26,] 0.4445355 0.5743531 0.7348531
#> [27,] 0.5738106 0.6163687 0.7701404
#> [28,] 0.5818155 0.6725833 0.8151138
#> [29,] 0.5552092 0.5941404 0.7822701
#> [30,] 0.5206797 0.6542805 0.7574722
#> [31,] 0.4608132 0.6472803 0.7070161
#> [32,] 0.5509821 0.6726182 0.7793215
#> [33,] 0.3533365 0.5435943 0.5249256
#> [34,] 0.4891743 0.5814191 0.7343600
#> [35,] 0.5030822 0.5843149 0.6900437
#> [36,] 0.5777037 0.6772789 0.7792009
#> [37,] 0.5155459 0.6610662 0.7568760
#> [38,] 0.5132004 0.6262293 0.7223906
#> [39,] 0.5136882 0.6069337 0.7423027
#> [40,] 0.3271486 0.5259861 0.6088212
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5945499 0.4054501
#>  [2,] 0.5880883 0.4119117
#>  [3,] 0.5820350 0.4179650
#>  [4,] 0.5898670 0.4101330
#>  [5,] 0.5859657 0.4140343
#>  [6,] 0.5792391 0.4207609
#>  [7,] 0.5896226 0.4103774
#>  [8,] 0.5882514 0.4117486
#>  [9,] 0.5890526 0.4109474
#> [10,] 0.5817904 0.4182096
#> [11,] 0.5827309 0.4172691
#> [12,] 0.5957415 0.4042585
#> [13,] 0.5859899 0.4140101
#> [14,] 0.5761554 0.4238446
#> [15,] 0.5679255 0.4320745
#> [16,] 0.5870541 0.4129459
#> [17,] 0.5891672 0.4108328
#> [18,] 0.5921746 0.4078254
#> [19,] 0.5771468 0.4228532
#> [20,] 0.5880701 0.4119299
#> [21,] 0.5893523 0.4106477
#> [22,] 0.5904537 0.4095463
#> [23,] 0.5794806 0.4205194
#> [24,] 0.5858437 0.4141563
#> [25,] 0.5871587 0.4128413
#> [26,] 0.5906998 0.4093002
#> [27,] 0.5879696 0.4120304
#> [28,] 0.5941571 0.4058429
#> [29,] 0.5896368 0.4103632
#> [30,] 0.5901025 0.4098975
#> [31,] 0.5891951 0.4108049
#> [32,] 0.5918599 0.4081401
#> [33,] 0.5763607 0.4236393
#> [34,] 0.5876879 0.4123121
#> [35,] 0.5868841 0.4131159
#> [36,] 0.5913648 0.4086352
#> [37,] 0.5904979 0.4095021
#> [38,] 0.5891765 0.4108235
#> [39,] 0.5874571 0.4125429
#> [40,] 0.5842562 0.4157438
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]        [,4]       [,5]         [,6]
#>  [1,] 2.2907313 1.5690333  1.41708317  1.30558930 -0.4031317 -0.109397627
#>  [2,] 2.1777079 1.6131460  0.80539978  0.73596925 -1.0553990 -0.199066848
#>  [3,] 1.5931202 1.0098302  0.62370842  0.34223836 -1.2925031 -0.420009981
#>  [4,] 2.3301740 1.7126529  1.06435418  0.80095141 -0.8233272 -0.171243243
#>  [5,] 2.1538053 1.6344617  0.83601761  0.70643282 -0.8718725  0.043265387
#>  [6,] 1.4350611 0.6419712 -0.03371215  0.03948106 -1.8822656 -1.055164942
#>  [7,] 2.4305499 1.9174636  1.19726789  1.01786085 -0.6704464  0.217387690
#>  [8,] 1.8410707 1.2925011  0.98379560  0.89614541 -0.5195301 -0.145059369
#>  [9,] 1.8420497 1.1836657  1.10485340  0.99612941 -0.5126297 -0.147681569
#> [10,] 1.8755360 1.3593433  0.47940202  0.19720772 -1.7586108 -0.451439624
#> [11,] 1.3396069 0.5626307  0.57634421  0.61384583 -0.7785681 -0.513645004
#> [12,] 2.5163373 1.8776769  1.55526180  1.36907311 -0.4979628 -0.113686813
#> [13,] 1.9878818 1.3580283  0.60987386  0.59283200 -1.0676590 -0.325331005
#> [14,] 1.2398203 0.5578412  0.19184852 -0.09511608 -1.7637201 -0.877850222
#> [15,] 0.9482722 0.5255147 -0.32060009 -0.39601527 -2.1041075 -0.880252226
#> [16,] 2.1272429 1.6387118  0.93334315  0.81645212 -0.7870089 -0.045811992
#> [17,] 2.1389676 1.7039575  1.12067643  0.86546752 -0.7874883 -0.094725377
#> [18,] 2.7311836 2.2011994  1.46528428  1.24680819 -0.4720766  0.307618626
#> [19,] 1.3632655 0.5887466 -0.10449017  0.03076799 -1.7549803 -0.871996387
#> [20,] 2.0827172 1.6537490  1.22784447  0.83819270 -0.7631269 -0.002018415
#> [21,] 2.4478310 2.0094539  1.24322933  1.01463221 -0.6326233  0.313978407
#> [22,] 2.0222882 1.3643643  1.01813979  1.00384777 -0.4645355 -0.202006530
#> [23,] 1.5217079 1.0635991  0.10374966  0.08480992 -1.7618724 -0.783994369
#> [24,] 1.8680914 1.4905302  1.07320211  0.81209879 -0.4975134  0.061801748
#> [25,] 2.1908440 1.7536256  1.18790506  0.88802238 -0.5921395  0.264662614
#> [26,] 2.1148544 1.5346730  1.05302573  0.74185210 -1.2818342 -0.514477935
#> [27,] 2.0243362 1.6885568  1.13680260  0.94892746 -0.4518174  0.128031387
#> [28,] 2.2194952 1.5472328  1.32692634  1.18782691 -0.5351200 -0.318339573
#> [29,] 2.1198951 1.7056729  1.33333734  0.97344676 -0.6522037  0.019783492
#> [30,] 2.4842913 1.8548588  1.18196259  0.95866746 -0.9079611  0.005837895
#> [31,] 2.5504811 1.8473165  0.92985721  0.78890573 -1.3193814 -0.154684455
#> [32,] 2.1203974 1.4533279  1.07828682  1.03929891 -0.6014160 -0.285746206
#> [33,] 1.6217597 1.0819798  0.14685509  0.09678395 -1.5557519 -0.270628892
#> [34,] 2.1006534 1.6749806  1.11863159  0.76438925 -0.9450193 -0.045348019
#> [35,] 1.7779563 1.0894941  0.91896671  0.75356656 -0.9796273 -0.291523278
#> [36,] 2.0552769 1.3691708  1.11541158  1.10340130 -0.4106784 -0.162111319
#> [37,] 2.3938557 1.8239327  1.13336779  0.96224917 -0.7716748 -0.038073676
#> [38,] 1.9438131 1.2563580  0.86632062  0.82804036 -0.8803734 -0.374400447
#> [39,] 2.2327409 1.7833547  1.17272882  0.83662219 -0.7744879  0.178409748
#> [40,] 1.8586249 1.1588710  0.62518578  0.28459284 -1.5771847 -0.548676646
#>              [,7]        [,8]        [,9]        [,10]
#>  [1,]  0.54247996  0.34110355  0.82684099  1.546891980
#>  [2,]  0.53867838 -0.15184569  0.43015682  0.919969032
#>  [3,]  0.13393143 -0.56044390  0.11812429  0.446175238
#>  [4,]  0.67328972 -0.01586842  0.45206521  1.145083645
#>  [5,]  0.74042547 -0.06281809  0.42394017  0.923578832
#>  [6,] -0.30513759 -0.94430172  0.04902329  0.058670539
#>  [7,]  0.88591484  0.14767753  0.67047090  1.214543498
#>  [8,]  0.51288073  0.13263704  0.45664921  1.097954073
#>  [9,]  0.54464576  0.24664439  0.47795122  1.182951932
#> [10,]  0.26738117 -0.68549284  0.03679321  0.280566068
#> [11,]  0.06817223 -0.17817382  0.39522407  0.592431408
#> [12,]  0.72129435  0.41365759  0.80251953  1.662290496
#> [13,]  0.44853177 -0.22529445  0.35271267  0.779472880
#> [14,] -0.24821844 -1.09211614 -0.20907838  0.001472854
#> [15,] -0.38231531 -1.13656230 -0.24323730 -0.423962904
#> [16,]  0.77933349  0.16023987  0.57809388  0.967817818
#> [17,]  0.68723124  0.04576761  0.44726738  1.101421025
#> [18,]  1.14324288  0.42770572  0.75608913  1.542146023
#> [19,] -0.24607754 -0.88148117  0.07979585  0.065452425
#> [20,]  0.76382350  0.11401073  0.37220909  1.137685185
#> [21,]  0.98177388  0.17580318  0.53759824  1.248694597
#> [22,]  0.51325970  0.21709324  0.67024019  1.161608381
#> [23,] -0.11966668 -0.88862841  0.14741819  0.138517980
#> [24,]  0.83149781  0.23850128  0.38986584  1.075197209
#> [25,]  0.95361791  0.18171735  0.47489557  1.130378335
#> [26,]  0.33162629 -0.22277459  0.29963428  1.019386471
#> [27,]  0.83359370  0.29741572  0.47416350  1.209104249
#> [28,]  0.53890203  0.33023047  0.71989212  1.483587165
#> [29,]  0.83515034  0.22174106  0.38110851  1.278944964
#> [30,]  0.87435842  0.08276590  0.63790837  1.138870560
#> [31,]  0.68835396 -0.15706930  0.60710567  0.880935539
#> [32,]  0.43589960  0.20463959  0.72005081  1.261716949
#> [33,]  0.22634550 -0.60440533  0.17482111  0.099785238
#> [34,]  0.66893752 -0.04330940  0.32860160  1.016857334
#> [35,]  0.24962769  0.01232901  0.34051194  0.800323530
#> [36,]  0.51315717  0.31335392  0.74129489  1.261015476
#> [37,]  0.75722732  0.06220356  0.66804933  1.135627854
#> [38,]  0.23236728  0.05281379  0.51607413  0.956350903
#> [39,]  0.86161201  0.05476640  0.43444146  1.057971323
#> [40,]  0.05901503 -0.72111000  0.10403834  0.442360010
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] 0.03534871 -0.3474581
#>  [2,] 0.02522844 -0.3308396
#>  [3,] 0.02860384 -0.3025289
#>  [4,] 0.02854315 -0.3348723
#>  [5,] 0.01634452 -0.3309679
#>  [6,] 0.03887113 -0.2807797
#>  [7,] 0.01553571 -0.3468698
#>  [8,] 0.02991423 -0.3268273
#>  [9,] 0.03049258 -0.3295576
#> [10,] 0.02549578 -0.3046315
#> [11,] 0.03275330 -0.3012407
#> [12,] 0.03412595 -0.3536264
#> [13,] 0.02691577 -0.3204966
#> [14,] 0.03536975 -0.2716408
#> [15,] 0.01987830 -0.2535140
#> [16,] 0.01773890 -0.3340617
#> [17,] 0.02544345 -0.3350804
#> [18,] 0.01609978 -0.3568627
#> [19,] 0.03120869 -0.2798630
#> [20,] 0.02181940 -0.3341734
#> [21,] 0.01430858 -0.3469800
#> [22,] 0.03277990 -0.3330612
#> [23,] 0.03046169 -0.2901799
#> [24,] 0.01888362 -0.3279261
#> [25,] 0.01351964 -0.3387123
#> [26,] 0.03813671 -0.3287224
#> [27,] 0.01991436 -0.3356637
#> [28,] 0.03869045 -0.3424874
#> [29,] 0.02413456 -0.3383295
#> [30,] 0.01909569 -0.3452933
#> [31,] 0.01989682 -0.3407423
#> [32,] 0.03472267 -0.3369370
#> [33,] 0.01176465 -0.2960865
#> [34,] 0.02208232 -0.3323330
#> [35,] 0.03017193 -0.3209271
#> [36,] 0.03275468 -0.3368557
#> [37,] 0.02322832 -0.3427958
#> [38,] 0.03410458 -0.3264575
#> [39,] 0.01545777 -0.3380051
#> [40,] 0.03458355 -0.3056868
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4054501 -0.4054501
#>  [2,]  0.4119117 -0.4119117
#>  [3,]  0.4179650 -0.4179650
#>  [4,] -0.5898670  0.5898670
#>  [5,] -0.5859657  0.5859657
#>  [6,]  0.4207609 -0.4207609
#>  [7,] -0.5896226  0.5896226
#>  [8,]  0.4117486 -0.4117486
#>  [9,] -0.5890526  0.5890526
#> [10,]  0.4182096 -0.4182096
#> [11,]  0.4172691 -0.4172691
#> [12,]  0.4042585 -0.4042585
#> [13,] -0.5859899  0.5859899
#> [14,]  0.4238446 -0.4238446
#> [15,] -0.5679255  0.5679255
#> [16,] -0.5870541  0.5870541
#> [17,]  0.4108328 -0.4108328
#> [18,] -0.5921746  0.5921746
#> [19,]  0.4228532 -0.4228532
#> [20,]  0.4119299 -0.4119299
#> [21,] -0.5893523  0.5893523
#> [22,] -0.5904537  0.5904537
#> [23,] -0.5794806  0.5794806
#> [24,] -0.5858437  0.5858437
#> [25,] -0.5871587  0.5871587
#> [26,]  0.4093002 -0.4093002
#> [27,] -0.5879696  0.5879696
#> [28,]  0.4058429 -0.4058429
#> [29,]  0.4103632 -0.4103632
#> [30,] -0.5901025  0.5901025
#> [31,] -0.5891951  0.5891951
#> [32,]  0.4081401 -0.4081401
#> [33,] -0.5763607  0.5763607
#> [34,] -0.5876879  0.5876879
#> [35,] -0.5868841  0.5868841
#> [36,]  0.4086352 -0.4086352
#> [37,]  0.4095021 -0.4095021
#> [38,] -0.5891765  0.5891765
#> [39,] -0.5874571  0.5874571
#> [40,]  0.4157438 -0.4157438
#> 
#> $L
#> [1] 0.6945448 0.6917097 0.6909292 0.6915256 0.6832342 0.7165318
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5217391