Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.3766323 -0.15265690 -0.09979695 -0.094070619 -0.03936559  0.004408326
#>  [2,] -0.2813754 -0.16510795 -0.15506839 -0.082085511 -0.05478620 -0.075669821
#>  [3,] -0.3839457 -0.15323649 -0.09023867 -0.125591813 -0.08489487 -0.082179986
#>  [4,] -0.2981366 -0.11790509 -0.15517623 -0.128450068 -0.17863579 -0.049698514
#>  [5,] -0.2454298 -0.05502706 -0.17272799 -0.095135860 -0.16224885 -0.095060183
#>  [6,] -0.3250960 -0.03930549 -0.07529693 -0.007283064  0.01922188 -0.076289547
#>  [7,] -0.3440212 -0.11544966 -0.19073193 -0.104215802 -0.08579966 -0.129646920
#>  [8,] -0.3695832 -0.08283509 -0.04913602 -0.117380060 -0.01874880 -0.099510156
#>  [9,] -0.3245372 -0.09600836 -0.13934485 -0.142271505 -0.05204612 -0.055642000
#> [10,] -0.2016613 -0.14269910 -0.01079346 -0.042950367 -0.03552392 -0.045152250
#>               V15          V16         V17         V18          V19         V2
#>  [1,] -0.09869145  0.053283689 -0.06024567 -0.05253293  0.134968307 -0.2901726
#>  [2,] -0.10859009 -0.003059654  0.05260156 -0.07411795  0.032511249 -0.3174794
#>  [3,] -0.10411849 -0.019342387 -0.08109143 -0.01508153 -0.030037657 -0.1882193
#>  [4,] -0.04706144 -0.066313140 -0.08380030 -0.04106759 -0.035608205 -0.2036627
#>  [5,] -0.12054980 -0.020602398 -0.02018752  0.08229439  0.105577164 -0.2158006
#>  [6,]  0.03424257 -0.062326104 -0.07143377  0.03589429  0.011181495 -0.4192668
#>  [7,] -0.01619691 -0.114508362 -0.06806456 -0.01928388 -0.097887808 -0.3451281
#>  [8,] -0.01076211  0.060326509  0.01147225 -0.03788620  0.001771125 -0.2307315
#>  [9,] -0.04875124  0.018557364 -0.05410056 -0.10573208  0.049198966 -0.2399408
#> [10,]  0.04008560 -0.001313382  0.09747905  0.14988937 -0.030975156 -0.2926436
#>                 V20          V21         V22        V23          V24
#>  [1,]  0.0658217860  0.047064514  0.16093961 0.05658429 -0.018757536
#>  [2,]  0.1645741916  0.153277652  0.10870474 0.10927972 -0.001308160
#>  [3,] -0.0640345357 -0.060005426 -0.05303594 0.04212179  0.020443622
#>  [4,] -0.0035127018  0.105126614  0.16852871 0.05147450  0.109276908
#>  [5,]  0.0350640902  0.077132180  0.05481598 0.06163438  0.210367911
#>  [6,]  0.0134176286  0.137432840  0.05145438 0.11740752  0.173597054
#>  [7,] -0.0667964998 -0.112595956 -0.05142244 0.11245287  0.097918562
#>  [8,] -0.0002420208  0.126039625  0.17742695 0.20191951  0.131352295
#>  [9,] -0.0690416554  0.001004074 -0.04501487 0.02920692  0.102755569
#> [10,]  0.0651335844  0.061608556 -0.01433854 0.02729659  0.004473445
#>                V25         V26        V27         V28         V29         V3
#>  [1,]  0.144795018 0.165192611 0.05858234  0.10175346 0.090004436 -0.2687543
#>  [2,] -0.023732650 0.104484804 0.13664889  0.17174204 0.132446984 -0.2779021
#>  [3,]  0.166738839 0.086399591 0.12808873  0.19423373 0.084828940 -0.1766887
#>  [4,]  0.180667027 0.026376891 0.15963397 -0.01322913 0.000556396 -0.2294079
#>  [5,]  0.114677713 0.096664048 0.17850034  0.14564476 0.051391022 -0.2432068
#>  [6,]  0.040595297 0.025487179 0.16496040  0.17727772 0.079953379 -0.3754085
#>  [7,]  0.196627266 0.009483513 0.02304038  0.05229233 0.091290652 -0.2359554
#>  [8,]  0.172450985 0.271811283 0.20858356  0.16360451 0.171731591 -0.2728823
#>  [9,]  0.089487826 0.111597153 0.18311307  0.14396568 0.024671594 -0.3757116
#> [10,]  0.009038518 0.087542075 0.11859930  0.02312000 0.081795580 -0.3376118
#>                V30         V31         V32         V33         V34         V35
#>  [1,]  0.082554349 -0.04697902 -0.06987501  0.07037349  0.07924657  0.03201289
#>  [2,]  0.201718726  0.06752589  0.09122905  0.05064953  0.03837141  0.02412047
#>  [3,]  0.077809233  0.19015728  0.07029734 -0.03257350  0.05424788  0.02027232
#>  [4,] -0.082401237 -0.02904933 -0.09100890 -0.19254012 -0.06112740 -0.21295854
#>  [5,]  0.008508929 -0.03411614 -0.02257674  0.02881081  0.02094002  0.07861353
#>  [6,]  0.030786261  0.11753002  0.05872361  0.05081990 -0.06999278 -0.04055313
#>  [7,]  0.008578771  0.05066229 -0.02430605 -0.12397080 -0.01812444 -0.06596114
#>  [8,]  0.015969794 -0.09406765 -0.14837181 -0.02779218 -0.03144990 -0.16064166
#>  [9,]  0.039871469 -0.05344357 -0.01079679 -0.19074106 -0.17307043 -0.03177473
#> [10,]  0.082477026  0.02842235  0.09529532  0.01176930 -0.01258776  0.02851261
#>                V36          V37           V38         V39         V4
#>  [1,]  0.094063275 -0.001686078 -0.0876099387 -0.10337664 -0.2786648
#>  [2,]  0.045118879 -0.032952951 -0.0431384529  0.06361264 -0.3491728
#>  [3,] -0.032755851 -0.106609226 -0.1361960343 -0.15507074 -0.3125644
#>  [4,] -0.244868062 -0.168175877 -0.0801583330 -0.15259541 -0.1882698
#>  [5,] -0.031509642 -0.018926102  0.0005827709 -0.02093061 -0.2861930
#>  [6,] -0.099570919  0.038351718  0.0190430780  0.04076872 -0.3511990
#>  [7,]  0.007496836 -0.082532749 -0.0555943448 -0.04502652 -0.1616805
#>  [8,] -0.226907922 -0.106924605 -0.1036692530 -0.13440744 -0.2762747
#>  [9,] -0.024454712 -0.105341786  0.0091047303 -0.11330995 -0.2827490
#> [10,]  0.127564187  0.108325082  0.0036261841 -0.06856550 -0.1894913
#>               V40          V41         V42         V43         V44         V45
#>  [1,]  0.04567629 -0.024922491 -0.09571997 -0.10951550 -0.03338484 -0.18351519
#>  [2,] -0.07363898 -0.027695922 -0.08354755 -0.04331432 -0.16440978 -0.16726116
#>  [3,] -0.02133792  0.003258388 -0.03212846 -0.14842541 -0.14747028 -0.02208615
#>  [4,] -0.09121224 -0.234338599 -0.04783094 -0.17195463 -0.23029189 -0.24104993
#>  [5,] -0.01771370 -0.019592529  0.04187898 -0.07635806 -0.14995926 -0.21792193
#>  [6,] -0.03554528 -0.030144987 -0.11744507 -0.14013992 -0.08598673 -0.07462015
#>  [7,] -0.12726560 -0.133137824 -0.14181580 -0.06034686 -0.17067327 -0.10804919
#>  [8,] -0.01586740 -0.093033651 -0.06583108 -0.19324058 -0.10772009 -0.18801038
#>  [9,] -0.01044509 -0.043964628 -0.09360109 -0.08330427 -0.07393393 -0.10712958
#> [10,] -0.03931052  0.063361635 -0.02474579 -0.01560515 -0.06718969 -0.13951949
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.18906414 -0.20964817 -0.2825960 -0.2559440 -0.2483959 -0.3424734
#>  [2,] -0.12513906 -0.20640740 -0.2246251 -0.2504045 -0.2006297 -0.4363225
#>  [3,] -0.16276511 -0.11492781 -0.1261663 -0.3320596 -0.1973619 -0.1969702
#>  [4,] -0.09525357 -0.19731628 -0.1822015 -0.2257842 -0.3099355 -0.1940550
#>  [5,] -0.16864890 -0.09840943 -0.1931048 -0.1943555 -0.1229999 -0.2573532
#>  [6,] -0.13679270 -0.18147215 -0.1873262 -0.4384462 -0.3197736 -0.4392397
#>  [7,] -0.15973850 -0.17726742 -0.3251734 -0.2874292 -0.2090860 -0.3652459
#>  [8,] -0.15728940 -0.30846789 -0.2202441 -0.2089514 -0.3239548 -0.2612238
#>  [9,] -0.15401481 -0.23253261 -0.1905832 -0.2914987 -0.2055564 -0.3471641
#> [10,] -0.17314444 -0.10374953 -0.1498996 -0.2448616 -0.2129760 -0.2717843
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3005813 -0.3947776 -0.2858757 -0.3327751 -0.4294797 -0.3022695
#>  [2,] -0.3362169 -0.3856924 -0.4131297 -0.4176332 -0.3598446 -0.3238556
#>  [3,] -0.2800802 -0.2821641 -0.3931470 -0.3390891 -0.3882159 -0.4140731
#>  [4,] -0.2997910 -0.2203615 -0.3217150 -0.3932904 -0.2926571 -0.4081032
#>  [5,] -0.2168307 -0.3133154 -0.2051939 -0.3884643 -0.2900522 -0.2929507
#>  [6,] -0.4095388 -0.4773694 -0.3413670 -0.4574240 -0.4066865 -0.4228950
#>  [7,] -0.3295900 -0.2590357 -0.2543814 -0.2011909 -0.4019389 -0.2654056
#>  [8,] -0.4408711 -0.2562177 -0.2845682 -0.3982226 -0.3516178 -0.3526946
#>  [9,] -0.3944712 -0.2941827 -0.3759327 -0.3195662 -0.3303929 -0.3351582
#> [10,] -0.2336685 -0.2360194 -0.3528594 -0.2102326 -0.2671722 -0.3095362
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.3051318 -0.3175868 -0.3881501 -0.09169789 -0.3794369 -0.12227766
#>  [2,] -0.3354764 -0.4723215 -0.3582987 -0.30281310 -0.4908437 -0.27420748
#>  [3,] -0.3328502 -0.2713061 -0.3505993 -0.22821909 -0.3417798 -0.20676453
#>  [4,] -0.2860127 -0.3531945 -0.2560537 -0.18643643 -0.2897856 -0.18497263
#>  [5,] -0.2229629 -0.2151049 -0.2288348 -0.18958660 -0.2463814 -0.21512661
#>  [6,] -0.3327508 -0.4210238 -0.4061607 -0.26391328 -0.3774544 -0.20377455
#>  [7,] -0.2289481 -0.3059867 -0.3012439 -0.22327672 -0.3141502 -0.21883427
#>  [8,] -0.4380934 -0.2451318 -0.3864057 -0.09634687 -0.3840343 -0.27684074
#>  [9,] -0.4413577 -0.4172907 -0.3140288 -0.27946438 -0.3827665 -0.10489234
#> [10,] -0.2635380 -0.3419116 -0.3508360 -0.16404271 -0.2503787 -0.07717862
#>                V8          V9
#>  [1,] -0.13333890 -0.05873166
#>  [2,] -0.08830394 -0.07813350
#>  [3,] -0.12540998 -0.15718813
#>  [4,] -0.17537314 -0.20818707
#>  [5,] -0.08916535 -0.18140045
#>  [6,] -0.28421815 -0.25280014
#>  [7,] -0.17705973 -0.08934745
#>  [8,] -0.11274111 -0.10090398
#>  [9,] -0.22242523 -0.11793828
#> [10,] -0.11842559 -0.10633311
#> 
#> $W[[2]]
#>              [,1]         [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,] -0.002901483  0.018093743 -0.000470464  0.05888207 -0.06133101 -0.05324173
#> [2,]  0.003435234 -0.008339542  0.047034056 -0.01448218 -0.03414387  0.02213550
#>             [,7]       [,8]        [,9]      [,10]
#> [1,] -0.02349443 0.01860361 -0.04342709 0.01659931
#> [2,]  0.04267966 0.01644466  0.04423443 0.06802187
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1           V10           V11           V12           V13
#>  [1,] 1.080702e-04  0.0004084951  0.0005633859  0.0007044789  8.277220e-04
#>  [2,] 7.870223e-05  0.0001172516  0.0002388774  0.0004202150  5.905922e-04
#>  [3,] 1.054378e-04  0.0005442918  0.0006841890  0.0007949996  9.174058e-04
#>  [4,] 6.894821e-06 -0.0005228840 -0.0005994071 -0.0004234090 -6.122579e-05
#>  [5,] 1.270473e-04  0.0006129147  0.0007880533  0.0009049283  1.007607e-03
#>  [6,] 1.932461e-04  0.0011904997  0.0014434336  0.0014727592  1.473762e-03
#>  [7,] 6.910295e-05  0.0004412259  0.0005256830  0.0005957413  6.938161e-04
#>  [8,] 1.358614e-04  0.0005408135  0.0007395037  0.0008701606  9.666922e-04
#>  [9,] 1.645623e-04  0.0011469623  0.0013887914  0.0014262853  1.406755e-03
#> [10,] 1.628917e-04  0.0009639879  0.0011824914  0.0012523538  1.291536e-03
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0009186388 0.0010038232 0.0011171175 0.0010941765 0.0012633692
#>  [2,] 0.0007532658 0.0008732235 0.0009804235 0.0009528237 0.0011453932
#>  [3,] 0.0009773462 0.0010415189 0.0011936143 0.0012150026 0.0013477737
#>  [4,] 0.0003727679 0.0005680148 0.0007333700 0.0005960792 0.0007818396
#>  [5,] 0.0010567463 0.0011213494 0.0012476117 0.0012293022 0.0013825817
#>  [6,] 0.0013772549 0.0013784022 0.0015087539 0.0014898723 0.0015921533
#>  [7,] 0.0007379924 0.0007820079 0.0009286809 0.0009849233 0.0010600910
#>  [8,] 0.0010176489 0.0010890223 0.0011737722 0.0011075282 0.0012981881
#>  [9,] 0.0012695728 0.0012396081 0.0013717111 0.0014102282 0.0015138441
#> [10,] 0.0012474934 0.0012642946 0.0013975718 0.0013959153 0.0015215419
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0015244067  1.173098e-04 0.0017857002 0.0018650521 0.0019165838
#>  [2,] 0.0012627407  6.915437e-05 0.0014268278 0.0014921565 0.0016086644
#>  [3,] 0.0016044238  1.274966e-04 0.0018700643 0.0019778824 0.0020389945
#>  [4,] 0.0005422163 -5.211183e-05 0.0006533593 0.0007239178 0.0009048471
#>  [5,] 0.0017343678  1.493122e-04 0.0020635768 0.0021553383 0.0021654562
#>  [6,] 0.0022468892  2.547043e-04 0.0027827498 0.0028777134 0.0027177065
#>  [7,] 0.0011929632  9.111352e-05 0.0013506834 0.0014628528 0.0015586049
#>  [8,] 0.0017141611  1.521901e-04 0.0020742122 0.0021136157 0.0020801917
#>  [9,] 0.0021118344  2.253246e-04 0.0025632864 0.0026831053 0.0026032889
#> [10,] 0.0020432177  2.077795e-04 0.0024840697 0.0025979168 0.0025327849
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.002037538 0.002245002 0.002501507 0.002523018 0.002354783 0.002333747
#>  [2,] 0.001811423 0.002132900 0.002517478 0.002462215 0.002216775 0.002139744
#>  [3,] 0.002123482 0.002261369 0.002430035 0.002491491 0.002385602 0.002388118
#>  [4,] 0.001321791 0.001964908 0.002722788 0.002607666 0.002224887 0.002026609
#>  [5,] 0.002233389 0.002373403 0.002553313 0.002624406 0.002500412 0.002510152
#>  [6,] 0.002607026 0.002506975 0.002434622 0.002644345 0.002651398 0.002758092
#>  [7,] 0.001638856 0.001737219 0.001829563 0.001889821 0.001849414 0.001864023
#>  [8,] 0.002149190 0.002320603 0.002560221 0.002604285 0.002421117 0.002414063
#>  [9,] 0.002500848 0.002378021 0.002253074 0.002448020 0.002492161 0.002582912
#> [10,] 0.002509299 0.002516180 0.002541266 0.002691111 0.002662123 0.002726750
#>               V29            V3         V30         V31         V32         V33
#>  [1,] 0.002286966  1.448456e-04 0.002062986 0.001754586 0.001430156 0.001478285
#>  [2,] 0.002158567  1.003247e-04 0.001977550 0.001796793 0.001388339 0.001526537
#>  [3,] 0.002290027  1.510085e-04 0.002062603 0.001741785 0.001463182 0.001434716
#>  [4,] 0.002127553 -1.815996e-05 0.002157472 0.002234534 0.001583648 0.001949470
#>  [5,] 0.002416795  1.740828e-04 0.002164667 0.001768307 0.001493710 0.001482617
#>  [6,] 0.002521345  2.717286e-04 0.002201069 0.001542960 0.001472065 0.001294139
#>  [7,] 0.001779911  1.077651e-04 0.001609211 0.001392149 0.001185386 0.001119186
#>  [8,] 0.002353701  1.795217e-04 0.002123414 0.001721857 0.001439004 0.001493386
#>  [9,] 0.002343333  2.422697e-04 0.001996628 0.001387171 0.001342728 0.001118584
#> [10,] 0.002532993  2.294697e-04 0.002215727 0.001670674 0.001504978 0.001377345
#>               V34          V35          V36          V37          V38
#>  [1,] 0.001361581 0.0011868505 0.0011500825 0.0010363727 0.0008860487
#>  [2,] 0.001431027 0.0012772309 0.0012699623 0.0011245180 0.0007977080
#>  [3,] 0.001333285 0.0012124738 0.0011931733 0.0010789327 0.0009742859
#>  [4,] 0.001955302 0.0018875322 0.0019885989 0.0017366715 0.0009352759
#>  [5,] 0.001352647 0.0011660973 0.0011107825 0.0010112305 0.0009645134
#>  [6,] 0.001127623 0.0008995060 0.0007618227 0.0007268068 0.0010585502
#>  [7,] 0.001059898 0.0010245758 0.0010270911 0.0009188506 0.0008098733
#>  [8,] 0.001358822 0.0011054479 0.0010102997 0.0009067491 0.0008663784
#>  [9,] 0.000947727 0.0007522726 0.0006455191 0.0006262491 0.0009506932
#> [10,] 0.001219274 0.0010124682 0.0009178998 0.0008635884 0.0010342197
#>                V39            V4          V40          V41          V42
#>  [1,] 0.0009511785  1.817936e-04 0.0009612549 0.0009297134 0.0009347409
#>  [2,] 0.0008128789  1.140342e-04 0.0009092753 0.0008612853 0.0007697577
#>  [3,] 0.0010262534  1.936278e-04 0.0010130931 0.0009619657 0.0009581873
#>  [4,] 0.0006810187 -6.227206e-05 0.0008438692 0.0007285028 0.0003244797
#>  [5,] 0.0010610555  2.261748e-04 0.0010231243 0.0009989404 0.0010588449
#>  [6,] 0.0012811835  3.778243e-04 0.0010810677 0.0011177093 0.0014035256
#>  [7,] 0.0008060574  1.341552e-04 0.0007994333 0.0007353093 0.0006981256
#>  [8,] 0.0010110613  2.347712e-04 0.0009785801 0.0009865019 0.0010743444
#>  [9,] 0.0011651803  3.370746e-04 0.0010080666 0.0010327036 0.0012858908
#> [10,] 0.0011995559  3.099564e-04 0.0010759647 0.0010783455 0.0012628241
#>                 V43           V44           V45           V46           V47
#>  [1,]  0.0007368874  0.0005153364  4.327226e-04  0.0004377677  0.0003364890
#>  [2,]  0.0004942171  0.0002473825  7.476577e-05  0.0001844187  0.0001742023
#>  [3,]  0.0008134877  0.0006270778  5.639110e-04  0.0005096849  0.0003846182
#>  [4,] -0.0001047261 -0.0003904286 -7.486822e-04 -0.0003954434 -0.0002308962
#>  [5,]  0.0009082995  0.0007026207  6.707852e-04  0.0006060369  0.0004459078
#>  [6,]  0.0014058480  0.0012410447  1.410574e-03  0.0011407376  0.0007840038
#>  [7,]  0.0006170351  0.0005006870  4.323191e-04  0.0003686025  0.0002799732
#>  [8,]  0.0008825562  0.0006459968  6.376034e-04  0.0006096715  0.0004434560
#>  [9,]  0.0013206607  0.0011958572  1.331759e-03  0.0010475920  0.0007322191
#> [10,]  0.0012052181  0.0010249986  1.100653e-03  0.0009122861  0.0006461741
#>                 V48           V49            V5           V50           V51
#>  [1,]  2.305095e-04  1.281652e-04  0.0002171345  5.846905e-05  4.104726e-05
#>  [2,]  9.228112e-05  4.432771e-05  0.0001258047  3.729314e-05  1.516665e-05
#>  [3,]  2.822981e-04  1.611601e-04  0.0002410416  6.694891e-05  5.131649e-05
#>  [4,] -2.393252e-04 -1.654772e-04 -0.0001298121 -1.593688e-07 -4.316274e-05
#>  [5,]  3.251996e-04  1.863551e-04  0.0002795327  7.367358e-05  5.865023e-05
#>  [6,]  6.104581e-04  3.625017e-04  0.0004759461  1.171423e-04  1.117304e-04
#>  [7,]  2.116449e-04  1.201980e-04  0.0001729879  4.996287e-05  3.886031e-05
#>  [8,]  3.128334e-04  1.783200e-04  0.0002784284  7.142381e-05  5.628845e-05
#>  [9,]  5.787502e-04  3.465393e-04  0.0004407647  1.067322e-04  1.045442e-04
#> [10,]  4.948953e-04  2.904472e-04  0.0003929906  9.878275e-05  8.960174e-05
#>                 V52          V53           V54          V55          V56
#>  [1,]  4.135488e-05 3.718857e-05  2.959576e-05 3.368329e-05 1.940761e-05
#>  [2,]  2.261623e-05 2.906540e-05  1.743243e-05 2.772257e-05 1.261735e-05
#>  [3,]  4.509092e-05 3.809013e-05  3.439741e-05 3.444067e-05 2.319522e-05
#>  [4,] -2.336062e-05 1.514080e-05 -2.406187e-06 1.716215e-05 4.898131e-06
#>  [5,]  5.369735e-05 4.253162e-05  3.807535e-05 3.828957e-05 2.463412e-05
#>  [6,]  9.369617e-05 5.832689e-05  6.214346e-05 5.126257e-05 3.814018e-05
#>  [7,]  3.183970e-05 2.749201e-05  2.609015e-05 2.433173e-05 1.826480e-05
#>  [8,]  5.466021e-05 4.297624e-05  3.735256e-05 4.006120e-05 2.324886e-05
#>  [9,]  8.498514e-05 5.138270e-05  5.618798e-05 4.429504e-05 3.480736e-05
#> [10,]  7.659334e-05 5.142520e-05  5.191364e-05 4.519585e-05 3.237617e-05
#>                V57           V58          V59            V6          V60
#>  [1,] 2.184471e-05  2.307530e-05 2.329678e-05  2.637440e-04 1.861692e-05
#>  [2,] 1.734848e-05  1.437140e-05 1.607344e-05  1.805204e-04 1.472420e-05
#>  [3,] 2.383079e-05  2.615125e-05 2.606223e-05  3.009320e-04 2.009600e-05
#>  [4,] 1.168246e-05 -4.608140e-06 5.891207e-07 -2.570605e-05 7.209473e-06
#>  [5,] 2.537228e-05  2.940334e-05 2.840249e-05  3.241804e-04 2.140401e-05
#>  [6,] 3.421316e-05  4.824547e-05 4.330946e-05  4.830249e-04 2.918937e-05
#>  [7,] 1.860594e-05  1.899959e-05 1.975387e-05  2.389868e-04 1.550230e-05
#>  [8,] 2.477730e-05  3.008436e-05 2.818780e-05  3.019054e-04 2.092231e-05
#>  [9,] 3.050213e-05  4.352940e-05 3.904020e-05  4.589297e-04 2.604128e-05
#> [10,] 3.063741e-05  3.987023e-05 3.684717e-05  4.217253e-04 2.594622e-05
#>                 V7            V8            V9
#>  [1,] 3.140575e-04  0.0003078503  3.438214e-04
#>  [2,] 2.288716e-04  0.0001514498  9.596401e-05
#>  [3,] 3.600656e-04  0.0003842056  4.564779e-04
#>  [4,] 5.214074e-05 -0.0001394309 -3.555647e-04
#>  [5,] 3.804949e-04  0.0004199419  5.172589e-04
#>  [6,] 5.431233e-04  0.0007355165  1.012468e-03
#>  [7,] 2.879136e-04  0.0003044491  3.655766e-04
#>  [8,] 3.475726e-04  0.0003814370  4.643461e-04
#>  [9,] 5.123493e-04  0.0006872720  9.538141e-04
#> [10,] 4.814232e-04  0.0006072359  8.124568e-04
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.0246985  0.02324947  0.03117647  0.02405545  0.02468081  0.02289282
#> [2,] -0.0246985 -0.02324947 -0.03117647 -0.02405545 -0.02468081 -0.02289282
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.03224715  0.01979501  0.02584361  0.02104397
#> [2,] -0.03224715 -0.01979501 -0.02584361 -0.02104397
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6436645 0.5153496 0.4277474 0.2680904 0.3472983 0.5604548 0.2460907
#>  [8] 0.5565778 0.4873526 0.5212454
#> 
#> $B[[2]]
#> [1]  0.1200981 -0.1432209
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.003512440 0.003404156 0.003480380 0.003464699 0.003667457 0.003766994
#>  [7] 0.002641145 0.003645264 0.003402082 0.003763163
#> 
#> $vB[[2]]
#> [1]  0.01192455 -0.01192455
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#>  [2,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#>  [3,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#>  [4,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#>  [5,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#>  [6,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#>  [7,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [8,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#>  [9,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [10,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [11,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [12,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [13,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [14,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [15,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [16,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [17,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [18,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [19,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [20,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [21,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [22,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [23,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [24,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [25,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [26,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [27,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [28,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [29,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [30,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [31,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [32,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [33,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [34,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [35,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [36,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [37,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [38,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [39,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [40,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#>  [2,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#>  [3,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#>  [4,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#>  [5,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#>  [6,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#>  [7,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [8,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#>  [9,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [10,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [11,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [12,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [13,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [14,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [15,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [16,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [17,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [18,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [19,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [20,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [21,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [22,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [23,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [24,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [25,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [26,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [27,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [28,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [29,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [30,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [31,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [32,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [33,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [34,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [35,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [36,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [37,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [38,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [39,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [40,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#>  [2,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#>  [3,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#>  [4,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#>  [5,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#>  [6,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#>  [7,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [8,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#>  [9,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [10,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [11,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [12,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [13,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [14,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [15,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [16,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [17,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [18,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [19,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [20,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [21,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [22,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [23,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [24,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [25,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [26,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [27,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [28,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [29,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [30,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [31,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [32,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [33,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [34,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [35,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [36,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [37,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [38,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [39,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [40,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#>  [2,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#>  [3,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#>  [4,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#>  [5,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#>  [6,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#>  [7,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [8,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#>  [9,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [10,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [11,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [12,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [13,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [14,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [15,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [16,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [17,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [18,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [19,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [20,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [21,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [22,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [23,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [24,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [25,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [26,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [27,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [28,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [29,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [30,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [31,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [32,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [33,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [34,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [35,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [36,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [37,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [38,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [39,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [40,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#>  [2,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#>  [3,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#>  [4,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#>  [5,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#>  [6,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#>  [7,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [8,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#>  [9,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [10,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [11,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [12,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [13,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [14,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [15,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [16,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [17,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [18,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [19,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [20,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [21,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [22,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [23,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [24,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [25,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [26,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [27,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [28,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [29,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [30,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [31,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [32,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [33,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [34,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [35,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [36,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [37,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [38,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [39,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [40,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#>  [2,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#>  [3,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#>  [4,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#>  [5,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#>  [6,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#>  [7,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [8,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#>  [9,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [10,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [11,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [12,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [13,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [14,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [15,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [16,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [17,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [18,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [19,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [20,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [21,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [22,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [23,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [24,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [25,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [26,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [27,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [28,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [29,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [30,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [31,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [32,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [33,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [34,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [35,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [36,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [37,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [38,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [39,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [40,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6638519 0.6644281 0.4907487 0.3325480 0.6019080 0.6533092 0.3291438
#>  [2,] 0.7240275 0.7305724 0.6273840 0.3800371 0.6685997 0.7190923 0.4508264
#>  [3,] 0.7142390 0.7270965 0.6580847 0.4888738 0.6825844 0.7323640 0.4953415
#>  [4,] 0.7075461 0.7348281 0.6001176 0.3654714 0.6478267 0.6887715 0.4612828
#>  [5,] 0.7217555 0.7418314 0.6594864 0.4508784 0.6688857 0.7154942 0.4934300
#>  [6,] 0.6504713 0.6809972 0.5910926 0.1970616 0.5842408 0.6320841 0.3597995
#>  [7,] 0.6006214 0.6210981 0.3852397 0.1687841 0.4991154 0.5032281 0.2393861
#>  [8,] 0.7538928 0.7442492 0.5852040 0.3414783 0.7013316 0.7174303 0.4313766
#>  [9,] 0.7110967 0.7044744 0.5688722 0.4534823 0.6667191 0.7007940 0.4150388
#> [10,] 0.7658823 0.7733887 0.6594796 0.5319679 0.7247594 0.7631348 0.4959477
#> [11,] 0.7538251 0.7712625 0.6761186 0.5104883 0.7160392 0.7634058 0.5157024
#> [12,] 0.7946086 0.8085375 0.6948674 0.4945184 0.7558282 0.7815761 0.5370690
#> [13,] 0.7028577 0.7051042 0.5302704 0.3965221 0.6421995 0.6662055 0.3489434
#> [14,] 0.5954594 0.6316600 0.4838902 0.2616465 0.5697696 0.5907910 0.2786026
#> [15,] 0.7281366 0.7230210 0.6317437 0.4668293 0.6611998 0.7209692 0.4896631
#> [16,] 0.6924938 0.6834999 0.5374688 0.3957337 0.6399996 0.7074929 0.3936293
#> [17,] 0.7608973 0.7824535 0.6956748 0.5791205 0.7371676 0.7783917 0.5632062
#> [18,] 0.7306532 0.7316804 0.5560894 0.3951641 0.6627729 0.6939929 0.4087741
#> [19,] 0.7541498 0.7645320 0.6477347 0.4351552 0.7033226 0.7372958 0.4756687
#> [20,] 0.6913591 0.6641979 0.4734892 0.4102835 0.6169230 0.6514054 0.3564403
#> [21,] 0.5903887 0.6079661 0.3892179 0.2108254 0.4921716 0.5213548 0.2627296
#> [22,] 0.7739441 0.7750108 0.6071063 0.3682170 0.7270646 0.7415427 0.4479218
#> [23,] 0.7127258 0.7260212 0.6512750 0.4329445 0.6640240 0.7123166 0.4656619
#> [24,] 0.7198793 0.7101449 0.5556555 0.4186888 0.6825088 0.6926358 0.3885208
#> [25,] 0.7401201 0.7238326 0.6137122 0.4098418 0.6946210 0.7023118 0.4473493
#> [26,] 0.7892495 0.8063767 0.6765922 0.4883733 0.7486601 0.7767794 0.5263389
#> [27,] 0.7252669 0.7118095 0.5319219 0.4742533 0.6775737 0.7087262 0.4031154
#> [28,] 0.5736829 0.6310306 0.5210833 0.2560152 0.5429005 0.5769374 0.3059051
#> [29,] 0.7256656 0.7059998 0.6792849 0.5085334 0.6782545 0.6975000 0.5381031
#> [30,] 0.7069904 0.7262292 0.6604553 0.4436724 0.6636441 0.7234817 0.4744931
#> [31,] 0.7576122 0.7468286 0.6015345 0.5576360 0.7228314 0.7428642 0.4576443
#> [32,] 0.7346681 0.7236412 0.6461627 0.5102603 0.7045180 0.7308008 0.5250068
#> [33,] 0.6647745 0.6991193 0.5403241 0.3111473 0.5986782 0.6863121 0.3384665
#> [34,] 0.6888758 0.6952944 0.6064532 0.2743575 0.6113802 0.6274393 0.4246736
#> [35,] 0.7127511 0.7614506 0.5660416 0.3909284 0.6712671 0.7226241 0.4045431
#> [36,] 0.7305998 0.7319133 0.6314276 0.4271141 0.6992659 0.7169947 0.4774030
#> [37,] 0.7441603 0.7473921 0.6050768 0.4904769 0.7087753 0.7403216 0.4474880
#> [38,] 0.7641986 0.7805990 0.6323102 0.5061184 0.7144111 0.7505287 0.4746885
#> [39,] 0.7162366 0.7062581 0.5590861 0.3965037 0.6658736 0.6828293 0.4095185
#> [40,] 0.7087156 0.7091323 0.5207247 0.3714913 0.6686321 0.6723543 0.3818870
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6834435 0.4842241 0.6333487
#>  [2,] 0.7133222 0.5377141 0.6965925
#>  [3,] 0.7761529 0.6039067 0.6855126
#>  [4,] 0.6182389 0.5352377 0.6989949
#>  [5,] 0.7212172 0.5892873 0.6910547
#>  [6,] 0.5098453 0.4580863 0.6964658
#>  [7,] 0.4079703 0.3398321 0.6265543
#>  [8,] 0.6846398 0.5550983 0.7301710
#>  [9,] 0.7565340 0.5625997 0.6857805
#> [10,] 0.8114773 0.6298087 0.7055484
#> [11,] 0.7850586 0.6394653 0.7223874
#> [12,] 0.7921685 0.6317072 0.7530619
#> [13,] 0.7022198 0.4850661 0.7016705
#> [14,] 0.5622821 0.4033233 0.6080012
#> [15,] 0.7563607 0.5848637 0.7048467
#> [16,] 0.7290389 0.5234732 0.6593029
#> [17,] 0.8105222 0.6593042 0.7205296
#> [18,] 0.6686357 0.5334689 0.7022647
#> [19,] 0.7507885 0.6091102 0.7360801
#> [20,] 0.7115398 0.4944274 0.6783919
#> [21,] 0.4472270 0.3648248 0.6164966
#> [22,] 0.7261793 0.5824135 0.7470026
#> [23,] 0.7276493 0.5817179 0.6905725
#> [24,] 0.7593825 0.5371299 0.7022751
#> [25,] 0.7008326 0.5412305 0.7117070
#> [26,] 0.7877433 0.6224777 0.7477883
#> [27,] 0.7692630 0.5605971 0.6926700
#> [28,] 0.5336430 0.4415057 0.6234693
#> [29,] 0.7469782 0.6463893 0.6962156
#> [30,] 0.7283893 0.5871636 0.6906999
#> [31,] 0.8090280 0.6154419 0.7226277
#> [32,] 0.7578565 0.6247331 0.6881457
#> [33,] 0.6631716 0.4891007 0.6706972
#> [34,] 0.4888149 0.4541975 0.6942915
#> [35,] 0.6929183 0.5335942 0.7200551
#> [36,] 0.6930044 0.5976710 0.7266969
#> [37,] 0.7996055 0.5867824 0.7128958
#> [38,] 0.7645858 0.6175188 0.7162575
#> [39,] 0.6786492 0.5545398 0.6961308
#> [40,] 0.6528700 0.5039272 0.7080083
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6012217 0.3987783
#>  [2,] 0.6043997 0.3956003
#>  [3,] 0.6075490 0.3924510
#>  [4,] 0.6029505 0.3970495
#>  [5,] 0.6063999 0.3936001
#>  [6,] 0.5961648 0.4038352
#>  [7,] 0.5937006 0.4062994
#>  [8,] 0.6033579 0.3966421
#>  [9,] 0.6061110 0.3938890
#> [10,] 0.6102246 0.3897754
#> [11,] 0.6090636 0.3909364
#> [12,] 0.6098756 0.3901244
#> [13,] 0.6044775 0.3955225
#> [14,] 0.5976828 0.4023172
#> [15,] 0.6067817 0.3932183
#> [16,] 0.6037067 0.3962933
#> [17,] 0.6114563 0.3885437
#> [18,] 0.6044149 0.3955851
#> [19,] 0.6066900 0.3933100
#> [20,] 0.6038759 0.3961241
#> [21,] 0.5946343 0.4053657
#> [22,] 0.6050567 0.3949433
#> [23,] 0.6056205 0.3943795
#> [24,] 0.6056062 0.3943938
#> [25,] 0.6053341 0.3946659
#> [26,] 0.6095880 0.3904120
#> [27,] 0.6069903 0.3930097
#> [28,] 0.5967060 0.4032940
#> [29,] 0.6074713 0.3925287
#> [30,] 0.6057848 0.3942152
#> [31,] 0.6104246 0.3895754
#> [32,] 0.6080243 0.3919757
#> [33,] 0.6008073 0.3991927
#> [34,] 0.5991049 0.4008951
#> [35,] 0.6047077 0.3952923
#> [36,] 0.6052532 0.3947468
#> [37,] 0.6084884 0.3915116
#> [38,] 0.6091474 0.3908526
#> [39,] 0.6039001 0.3960999
#> [40,] 0.6032610 0.3967390
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]        [,4]         [,5]       [,6]
#>  [1,] 0.6805070 0.6830906 -0.03700926 -0.69668320  0.413421637 0.63361700
#>  [2,] 0.9645284 0.9975285  0.52101029 -0.48939081  0.701858317 0.93996360
#>  [3,] 0.9160618 0.9799409  0.65477045 -0.04451197  0.765674271 1.00665002
#>  [4,] 0.8834959 1.0192582  0.40595502 -0.55169412  0.609500026 0.79438219
#>  [5,] 0.9531864 1.0555094  0.66100649 -0.19712228  0.703149510 0.92221992
#>  [6,] 0.6211114 0.7583581  0.36848385 -1.40476163  0.340207004 0.54116787
#>  [7,] 0.4080549 0.4942119 -0.46736650 -1.59426911 -0.003538579 0.01291244
#>  [8,] 1.1194830 1.0681725  0.34417356 -0.65671320  0.853646847 0.93175033
#>  [9,] 0.9007163 0.8686967  0.27725130 -0.18661050  0.693383045 0.85108170
#> [10,] 1.1852044 1.2275458  0.66097583  0.12804616  0.968193938 1.16994336
#> [11,] 1.1191182 1.2154537  0.73599107  0.04195932  0.924898475 1.17144309
#> [12,] 1.3529320 1.4405348  0.82297452 -0.02192725  1.129941772 1.27487473
#> [13,] 0.8609435 0.8717234  0.12122988 -0.41997762  0.584923411 0.69107250
#> [14,] 0.3865810 0.5393449 -0.06446139 -1.03742843  0.280911011 0.36723614
#> [15,] 0.9851887 0.9594965  0.53970472 -0.13287815  0.668645589 0.94927412
#> [16,] 0.8118043 0.7699029  0.15015673 -0.42327335  0.575362374 0.88323865
#> [17,] 1.1576055 1.2800220  0.82678550  0.31916413  1.031298890 1.25631839
#> [18,] 0.9979393 1.0031650  0.22530595 -0.42565607  0.675675985 0.81885339
#> [19,] 1.1208688 1.1776891  0.60909654 -0.26084813  0.863170113 1.03196036
#> [20,] 0.8064809 0.6820580 -0.10614265 -0.36279373  0.476508030 0.62522251
#> [21,] 0.3655727 0.4387707 -0.45060102 -1.31995699 -0.031316178 0.08547105
#> [22,] 1.2307176 1.2368244  0.43516481 -0.53987359  0.979780089 1.05400242
#> [23,] 0.9086598 0.9745286  0.62464837 -0.26984774  0.681278593 0.90666180
#> [24,] 0.9438631 0.8960877  0.22354837 -0.32815817  0.765325124 0.81247106
#> [25,] 1.0465926 0.9635528  0.46294352 -0.36461958  0.821812749 0.85833083
#> [26,] 1.3204077 1.4266365  0.73815475 -0.04651519  1.091479072 1.24699614
#> [27,] 0.9707398 0.9041888  0.12786170 -0.10307804  0.742643664 0.88920573
#> [28,] 0.2968935 0.5366408  0.08438334 -1.06678400  0.172025073 0.31021380
#> [29,] 0.9727415 0.8760346  0.75048745  0.03413695  0.745761593 0.83542124
#> [30,] 0.8808119 0.9755746  0.66532400 -0.22627099  0.679576088 0.96179844
#> [31,] 1.1396327 1.0817689  0.41186309  0.23157326  0.958549911 1.06090877
#> [32,] 1.0184373 0.9625958  0.60221396  0.04104706  0.868905807 0.99868914
#> [33,] 0.6846443 0.8431076  0.16164757 -0.79476080  0.399960830 0.78293373
#> [34,] 0.7948686 0.8249893  0.43242747 -0.97262548  0.453117526 0.52124679
#> [35,] 0.9087834 1.1606487  0.26571892 -0.44341135  0.713921682 0.95751546
#> [36,] 0.9976680 1.0043517  0.53834622 -0.29363518  0.843804376 0.92960266
#> [37,] 1.0677054 1.0847514  0.42666409 -0.03809699  0.889443606 1.04764070
#> [38,] 1.1758378 1.2691602  0.54214043  0.02447476  0.916905131 1.10143409
#> [39,] 0.9258696 0.8772793  0.23745394 -0.42005448  0.689580478 0.76680463
#> [40,] 0.8891544 0.8911737  0.08294614 -0.52582446  0.702004816 0.71885250
#>              [,7]        [,8]        [,9]     [,10]
#>  [1,] -0.71205989  0.76964229 -0.06312440 0.5466099
#>  [2,] -0.19733219  0.91157435  0.15114364 0.8311237
#>  [3,] -0.01863468  1.24338638  0.42177003 0.7792229
#>  [4,] -0.15517936  0.48208004  0.14118494 0.8425161
#>  [5,] -0.02628143  0.95050751  0.36101976 0.8050545
#>  [6,] -0.57623439  0.03938643 -0.16804912 0.8305245
#>  [7,] -1.15604791 -0.37236245 -0.66404248 0.5174629
#>  [8,] -0.27623689  0.77517756  0.22129193 0.9954902
#>  [9,] -0.34317335  1.13376994  0.25171957 0.7804657
#> [10,] -0.01620965  1.45963776  0.53139618 0.8738607
#> [11,]  0.06283034  1.29539307  0.57304433 0.9563350
#> [12,]  0.14854840  1.33804668  0.53954760 1.1150097
#> [13,] -0.62368672  0.85789070 -0.05975335 0.8552652
#> [14,] -0.95140395  0.25042925 -0.39163704 0.4389182
#> [15,] -0.04135334  1.13282957  0.34277192 0.8704856
#> [16,] -0.43208209  0.98975165  0.09396169 0.6601893
#> [17,]  0.25418471  1.45340707  0.66019510 0.9470901
#> [18,] -0.36903544  0.70202101  0.13407607 0.8581058
#> [19,] -0.09740216  1.10282225  0.44357333 1.0256935
#> [20,] -0.59084786  0.90287404 -0.02229137 0.7463915
#> [21,] -1.03182923 -0.21188109 -0.55448353 0.4747043
#> [22,] -0.20907093  0.97532346  0.33268910 1.0826897
#> [23,] -0.13756892  0.98272838  0.32982972 0.8027973
#> [24,] -0.45353435  1.14929690  0.14879345 0.8581551
#> [25,] -0.21138633  0.85126598  0.16529727 0.9036888
#> [26,]  0.10545335  1.31137611  0.50007819 1.0868513
#> [27,] -0.39250102  1.20415430  0.24358574 0.8126316
#> [28,] -0.81933354  0.13477558 -0.23505356 0.5042998
#> [29,]  0.15270833  1.08256058  0.60320518 0.8293412
#> [30,] -0.10211629  0.98646565  0.35225200 0.8033936
#> [31,] -0.16982986  1.44370667  0.47024548 0.9575337
#> [32,]  0.10011090  1.14096363  0.50968716 0.7914643
#> [33,] -0.67013565  0.67746032 -0.04360408 0.7113399
#> [34,] -0.30361654 -0.04474785 -0.18372490 0.8202599
#> [35,] -0.38657082  0.81379834  0.13457938 0.9447348
#> [36,] -0.09044976  0.81420272  0.39577025 0.9779279
#> [37,] -0.21082544  1.38383048  0.35067971 0.9094899
#> [38,] -0.10133271  1.17798785  0.47903012 0.9259725
#> [39,] -0.36595629  0.74757097  0.21903081 0.8289400
#> [40,] -0.48154635  0.63167866  0.01570916 0.8857306
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2191758 -0.1913823
#>  [2,] 0.2305680 -0.1932635
#>  [3,] 0.2404862 -0.1965350
#>  [4,] 0.2265954 -0.1911787
#>  [5,] 0.2375046 -0.1946996
#>  [6,] 0.2091430 -0.1803675
#>  [7,] 0.1984261 -0.1808589
#>  [8,] 0.2258288 -0.1936473
#>  [9,] 0.2346276 -0.1963663
#> [10,] 0.2463236 -0.2019327
#> [11,] 0.2446362 -0.1987415
#> [12,] 0.2460589 -0.2007304
#> [13,] 0.2298103 -0.1943465
#> [14,] 0.2086881 -0.1871314
#> [15,] 0.2393660 -0.1944382
#> [16,] 0.2268252 -0.1941086
#> [17,] 0.2512914 -0.2021466
#> [18,] 0.2290031 -0.1948917
#> [19,] 0.2378869 -0.1955328
#> [20,] 0.2280245 -0.1936169
#> [21,] 0.2018671 -0.1812897
#> [22,] 0.2306461 -0.1959340
#> [23,] 0.2353081 -0.1936320
#> [24,] 0.2323769 -0.1965033
#> [25,] 0.2322449 -0.1954959
#> [26,] 0.2448901 -0.2006904
#> [27,] 0.2358682 -0.1988105
#> [28,] 0.2093177 -0.1824410
#> [29,] 0.2422713 -0.1944239
#> [30,] 0.2360531 -0.1935748
#> [31,] 0.2465469 -0.2025504
#> [32,] 0.2407299 -0.1982850
#> [33,] 0.2201176 -0.1887125
#> [34,] 0.2163467 -0.1853903
#> [35,] 0.2301806 -0.1949392
#> [36,] 0.2332665 -0.1941358
#> [37,] 0.2410950 -0.1998677
#> [38,] 0.2430013 -0.2007287
#> [39,] 0.2278558 -0.1938866
#> [40,] 0.2248804 -0.1941909
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3987783 -0.3987783
#>  [2,] -0.6043997  0.6043997
#>  [3,]  0.3924510 -0.3924510
#>  [4,] -0.6029505  0.6029505
#>  [5,] -0.6063999  0.6063999
#>  [6,]  0.4038352 -0.4038352
#>  [7,]  0.4062994 -0.4062994
#>  [8,]  0.3966421 -0.3966421
#>  [9,]  0.3938890 -0.3938890
#> [10,]  0.3897754 -0.3897754
#> [11,] -0.6090636  0.6090636
#> [12,] -0.6098756  0.6098756
#> [13,] -0.6044775  0.6044775
#> [14,]  0.4023172 -0.4023172
#> [15,] -0.6067817  0.6067817
#> [16,]  0.3962933 -0.3962933
#> [17,] -0.6114563  0.6114563
#> [18,] -0.6044149  0.6044149
#> [19,] -0.6066900  0.6066900
#> [20,]  0.3961241 -0.3961241
#> [21,]  0.4053657 -0.4053657
#> [22,]  0.3949433 -0.3949433
#> [23,]  0.3943795 -0.3943795
#> [24,]  0.3943938 -0.3943938
#> [25,] -0.6053341  0.6053341
#> [26,] -0.6095880  0.6095880
#> [27,]  0.3930097 -0.3930097
#> [28,]  0.4032940 -0.4032940
#> [29,] -0.6074713  0.6074713
#> [30,]  0.3942152 -0.3942152
#> [31,] -0.6104246  0.6104246
#> [32,] -0.6080243  0.6080243
#> [33,]  0.3991927 -0.3991927
#> [34,] -0.5991049  0.5991049
#> [35,]  0.3952923 -0.3952923
#> [36,] -0.6052532  0.6052532
#> [37,]  0.3915116 -0.3915116
#> [38,] -0.6091474  0.6091474
#> [39,] -0.6039001  0.6039001
#> [40,]  0.3967390 -0.3967390
#> 
#> $L
#> [1] 0.6877020 0.7377836 1.1794917 0.6999095 0.7279377 0.7087659
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4492754