Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12          V13          V14
#>  [1,] -0.2642560 -0.07595060 -0.128286120 -0.12253545 -0.161331779  0.006123899
#>  [2,] -0.2518428 -0.11914554 -0.057145151 -0.06765589 -0.093595284  0.009191678
#>  [3,] -0.2396946 -0.17445967 -0.089328153 -0.13905512  0.009665242 -0.039695118
#>  [4,] -0.3696500 -0.16010186 -0.072626213 -0.01521637 -0.068224208  0.019396667
#>  [5,] -0.3808429 -0.18354431 -0.182842838 -0.11362225 -0.049303556 -0.017360919
#>  [6,] -0.2392074 -0.02270998  0.005141047 -0.09839286 -0.023493344 -0.062374073
#>  [7,] -0.2873271 -0.17418034 -0.069397349 -0.01231189  0.006444252 -0.091025726
#>  [8,] -0.2105913 -0.19550033 -0.027942027 -0.04096798 -0.160506960 -0.225993330
#>  [9,] -0.2099664 -0.08419640 -0.095212300 -0.04991592 -0.150045922 -0.165283530
#> [10,] -0.4149740 -0.06298544 -0.143063483 -0.06542007 -0.053453833 -0.127778048
#>               V15          V16         V17         V18          V19         V2
#>  [1,] -0.06702080  0.058668736  0.07021158  0.13104224  0.185910398 -0.2285121
#>  [2,] -0.11558009 -0.050150171 -0.03463039  0.01928735 -0.011926837 -0.3014111
#>  [3,] -0.01489693  0.006198172 -0.07597103 -0.05697651 -0.137287567 -0.2696371
#>  [4,]  0.07094850  0.064873519  0.02403851  0.10349546  0.062949916 -0.4053087
#>  [5,] -0.03483869  0.013883300  0.03558679 -0.05308088  0.130188888 -0.2072390
#>  [6,] -0.06140386 -0.099278317 -0.09231993  0.01658092 -0.011097136 -0.3904260
#>  [7,] -0.09009190  0.027091468 -0.05209975  0.01741907 -0.068177317 -0.2345942
#>  [8,] -0.18810969 -0.193130963 -0.02185171 -0.01699090  0.047322738 -0.3562876
#>  [9,] -0.12918179 -0.129040131 -0.08571513 -0.08381383  0.005559922 -0.2782157
#> [10,] -0.09458510 -0.015574630 -0.08119370 -0.05098342 -0.051431313 -0.2017040
#>                V20          V21          V22         V23          V24
#>  [1,]  0.186731970  0.110843995  0.111417352  0.23296868  0.108576238
#>  [2,]  0.046034680  0.120296583  0.097696920  0.20204351  0.216415833
#>  [3,] -0.085178585 -0.020207338 -0.022366867 -0.08979314 -0.078701170
#>  [4,]  0.246153160  0.302627974  0.302759833  0.12722777  0.230756108
#>  [5,] -0.012146887  0.141668047  0.041710562  0.06730295  0.138368815
#>  [6,] -0.059426917  0.013854225 -0.014757652  0.11303053 -0.005419794
#>  [7,] -0.045145603  0.021242504  0.180653939  0.14009817  0.069457142
#>  [8,] -0.050356967  0.002062944  0.064559124 -0.03683523  0.101449373
#>  [9,] -0.100111264 -0.091089817  0.006444383 -0.03115278  0.086866419
#> [10,]  0.009114184  0.100972312  0.052147351  0.15794203  0.176346194
#>              V25        V26        V27        V28         V29         V3
#>  [1,] 0.11178378 0.16885511 0.09127331 0.11335815  0.05542372 -0.3121260
#>  [2,] 0.06794868 0.15014997 0.22847858 0.14824649  0.07333702 -0.2699658
#>  [3,] 0.04636738 0.12383379 0.12233654 0.03699294  0.07164482 -0.3134915
#>  [4,] 0.26313581 0.32539303 0.31908596 0.27694197  0.20771194 -0.2921904
#>  [5,] 0.01485563 0.13477701 0.04055336 0.16610484  0.08577114 -0.3592357
#>  [6,] 0.15520786 0.15258956 0.13579321 0.10745024  0.08757717 -0.3027829
#>  [7,] 0.13930367 0.12467914 0.21087116 0.19751684  0.21744220 -0.2296525
#>  [8,] 0.18056971 0.02287145 0.08495644 0.04497094  0.14190515 -0.2394526
#>  [9,] 0.04213347 0.08231765 0.00362152 0.12473477 -0.07846930 -0.1910419
#> [10,] 0.19804944 0.10001242 0.14969502 0.09391492  0.18870161 -0.2433368
#>               V30          V31         V32          V33         V34         V35
#>  [1,]  0.05218158  0.102873185 -0.06076722 -0.007164902 -0.02717009  0.01456530
#>  [2,]  0.12556101 -0.041103617 -0.01680116 -0.123298696 -0.08296240 -0.02428998
#>  [3,]  0.03687752  0.095410578  0.01206274 -0.035728753 -0.14402393 -0.12790280
#>  [4,]  0.26108125  0.199875275  0.17008420  0.019619744  0.01323571 -0.01418250
#>  [5,]  0.03197552 -0.060136036  0.04095313  0.067980553  0.09801986  0.01562582
#>  [6,] -0.02957298 -0.071285402  0.07601436  0.052334907  0.03577570 -0.08208593
#>  [7,]  0.11696261  0.074471574 -0.07225022  0.066269128  0.04834419  0.05588673
#>  [8,]  0.02202596 -0.003054258  0.08001578  0.059159673  0.04460471 -0.06113608
#>  [9,]  0.02863787 -0.123226747 -0.07165839 -0.073015871 -0.22273242 -0.06781380
#> [10,]  0.10465626  0.092039365 -0.02674513  0.077110135  0.04421816 -0.07805144
#>                V36         V37         V38         V39         V4          V40
#>  [1,]  0.002673021  0.03445641  0.01520632 -0.11892163 -0.2141519 -0.134004911
#>  [2,] -0.039973759  0.02117649 -0.07410736 -0.09619813 -0.2789235 -0.110814744
#>  [3,] -0.078126884 -0.03163516 -0.03087083 -0.08207762 -0.1679362 -0.170545665
#>  [4,]  0.125196386  0.08654124  0.01848929  0.03847401 -0.3494410 -0.011354960
#>  [5,] -0.123084570 -0.14756852 -0.12561120  0.03594876 -0.3402722  0.027512874
#>  [6,] -0.076427536  0.01489295 -0.05733679 -0.05441989 -0.2096510  0.032947393
#>  [7,] -0.021174741 -0.09353465 -0.10343001 -0.14234025 -0.2250008 -0.009675006
#>  [8,] -0.020451176 -0.12970276 -0.13058576 -0.13848481 -0.2953610  0.008668150
#>  [9,] -0.033394337 -0.01473199 -0.06085667 -0.07955947 -0.2137892 -0.087671469
#> [10,] -0.143350067 -0.14048788 -0.02965732 -0.02157174 -0.2730337 -0.139330903
#>                V41          V42         V43         V44         V45         V46
#>  [1,] -0.147874825 -0.005901273 -0.08453392 -0.22005500 -0.19434974 -0.24159527
#>  [2,]  0.012822780 -0.050897939 -0.07951693 -0.05264546 -0.07935823 -0.12246454
#>  [3,] -0.076680477 -0.146315487 -0.09469527 -0.21463236 -0.08055706 -0.23830386
#>  [4,] -0.028595649 -0.027919404  0.01697197 -0.17249785 -0.08455551 -0.09124220
#>  [5,] -0.148575500 -0.080760814 -0.18341199 -0.07400345 -0.10109711 -0.17422100
#>  [6,]  0.001744561 -0.116448852 -0.08297331 -0.02546578 -0.15353327 -0.17921299
#>  [7,]  0.018409603 -0.034199667 -0.15964688 -0.20982811 -0.14687170 -0.13598500
#>  [8,] -0.170975942 -0.108820468 -0.04156642 -0.05913617 -0.20871530 -0.23971090
#>  [9,] -0.119927522 -0.059674646 -0.14199935 -0.14406141 -0.03767402 -0.14344307
#> [10,] -0.106439603 -0.034151200 -0.14578115 -0.04993250 -0.17395627 -0.07939715
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.1586512 -0.1712310 -0.2198279 -0.2821044 -0.2592431 -0.2444539
#>  [2,] -0.2204611 -0.3154608 -0.2276045 -0.1476540 -0.3534776 -0.3699748
#>  [3,] -0.2118294 -0.1178768 -0.2721073 -0.2063901 -0.3555526 -0.3087808
#>  [4,] -0.2219873 -0.2881430 -0.2867021 -0.3238179 -0.3708565 -0.4711032
#>  [5,] -0.1570628 -0.2690392 -0.3145145 -0.2942214 -0.3411176 -0.4023500
#>  [6,] -0.2631090 -0.2845453 -0.1995179 -0.2948586 -0.3926926 -0.2660268
#>  [7,] -0.1873128 -0.1715516 -0.3660264 -0.2761291 -0.3535303 -0.3036489
#>  [8,] -0.1168365 -0.2603470 -0.2844322 -0.1468803 -0.3363514 -0.2925181
#>  [9,] -0.1157096 -0.2858934 -0.2704202 -0.1572945 -0.2665471 -0.2407562
#> [10,] -0.1903385 -0.1330529 -0.3066958 -0.3125864 -0.3776248 -0.3218214
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.2962494 -0.3234466 -0.2494094 -0.3089449 -0.3230349 -0.2532921
#>  [2,] -0.2481957 -0.2814789 -0.2061666 -0.3963122 -0.2532345 -0.3225836
#>  [3,] -0.2730105 -0.3270623 -0.2715032 -0.2551862 -0.3424669 -0.3323578
#>  [4,] -0.3949638 -0.3992888 -0.4706928 -0.4718599 -0.3822054 -0.3569404
#>  [5,] -0.3599587 -0.3924006 -0.2315580 -0.3696972 -0.2700818 -0.4067554
#>  [6,] -0.3598522 -0.3704407 -0.3886238 -0.3141846 -0.4633650 -0.4516132
#>  [7,] -0.3818624 -0.3470241 -0.3391491 -0.4085050 -0.4050646 -0.4559471
#>  [8,] -0.2422293 -0.3312296 -0.3377253 -0.2464223 -0.2493035 -0.3249023
#>  [9,] -0.2907371 -0.2938742 -0.3516602 -0.1983679 -0.2217272 -0.2213908
#> [10,] -0.3419759 -0.2944659 -0.4145464 -0.2988897 -0.3489084 -0.3109117
#>              V58        V59         V6        V60          V7          V8
#>  [1,] -0.2782326 -0.3332158 -0.2774299 -0.3813766 -0.12743581 -0.13175689
#>  [2,] -0.2724768 -0.3806309 -0.1399020 -0.3930060 -0.20829701 -0.21014299
#>  [3,] -0.3386405 -0.2790141 -0.1070046 -0.1935387 -0.20978928 -0.14943615
#>  [4,] -0.3789091 -0.4654286 -0.2485873 -0.3816902 -0.27610869 -0.18609251
#>  [5,] -0.3209523 -0.3224609 -0.2510449 -0.3551554 -0.21578028 -0.14996015
#>  [6,] -0.3014507 -0.4309784 -0.2465639 -0.2771474 -0.19871072 -0.12554764
#>  [7,] -0.4330122 -0.3661172 -0.2486170 -0.3930376 -0.17878038 -0.15037036
#>  [8,] -0.3318308 -0.3285328 -0.1916162 -0.3935656 -0.22841262 -0.28165379
#>  [9,] -0.2502466 -0.3404962 -0.2604492 -0.3002470 -0.06793356 -0.09511829
#> [10,] -0.3136963 -0.3261256 -0.1226766 -0.4339273 -0.14763099 -0.15421624
#>                V9
#>  [1,] -0.10977065
#>  [2,] -0.13319536
#>  [3,] -0.06249894
#>  [4,] -0.20870638
#>  [5,] -0.16363407
#>  [6,] -0.13183736
#>  [7,] -0.18633868
#>  [8,] -0.06257929
#>  [9,] -0.12187423
#> [10,] -0.11480694
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] 0.05639707 -0.01446192 -0.04087305  0.1212199 -0.04168661  0.06354581
#> [2,] 0.01786907  0.04019094 -0.01827646 -0.1191075  0.00481514 -0.06668688
#>             [,7]        [,8]        [,9]        [,10]
#> [1,] -0.05529111  0.02635256  0.07918024 -0.003226509
#> [2,] -0.03477904 -0.08213484 -0.02467957 -0.083532835
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -6.644803e-06 -0.0000376777 -5.368628e-05 -5.695506e-05 -4.667209e-05
#>  [2,]  1.560908e-04  0.0008910719  1.195045e-03  1.219218e-03  9.150078e-04
#>  [3,]  3.160622e-05  0.0002000491  2.850525e-04  3.068516e-04  2.395206e-04
#>  [4,] -9.525045e-05 -0.0004756705 -6.135554e-04 -6.034480e-04 -4.975204e-04
#>  [5,]  1.103416e-04  0.0006591291  9.000763e-04  9.403112e-04  7.149293e-04
#>  [6,] -1.694577e-04 -0.0010235335 -1.406564e-03 -1.469171e-03 -1.100458e-03
#>  [7,]  8.641176e-05  0.0004942840  6.675645e-04  6.897901e-04  5.271142e-04
#>  [8,] -1.437738e-04 -0.0009146205 -1.267352e-03 -1.337731e-03 -9.844354e-04
#>  [9,] -1.169601e-04 -0.0007529571 -1.065943e-03 -1.140549e-03 -8.065208e-04
#> [10,] -4.580484e-05 -0.0002593786 -3.603768e-04 -3.766220e-04 -2.959828e-04
#>                 V14           V15           V16           V17           V18
#>  [1,] -2.112563e-05 -9.540536e-06 -7.407388e-06 -3.692576e-06 -4.909038e-06
#>  [2,]  5.166710e-04  2.563730e-04  1.145088e-04 -2.697466e-05  1.096773e-04
#>  [3,]  1.231258e-04  6.989333e-05  5.897225e-05  2.984352e-05  5.615110e-05
#>  [4,] -3.550310e-04 -2.593426e-04 -1.971175e-04 -1.571856e-04 -1.912720e-04
#>  [5,]  3.967430e-04  2.117162e-04  1.270573e-04  2.940500e-05  1.398617e-04
#>  [6,] -5.823608e-04 -2.785088e-04 -1.312492e-04  2.721289e-05 -1.497819e-04
#>  [7,]  3.093973e-04  1.733577e-04  1.053087e-04  3.736071e-05  1.150768e-04
#>  [8,] -4.992483e-04 -2.242599e-04 -8.430913e-05  5.604292e-05 -1.145420e-04
#>  [9,] -3.660145e-04 -1.297523e-04  3.606143e-08  1.322739e-04 -2.742319e-05
#> [10,] -1.570553e-04 -8.399972e-05 -5.526871e-05 -1.930252e-05 -5.035940e-05
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0000344114 -6.791115e-06 -4.496572e-05 -3.252532e-05  1.101420e-06
#>  [2,]  0.0007903630  1.544428e-04  1.331789e-03  1.319240e-03  8.013936e-04
#>  [3,]  0.0002104644  2.941078e-05  2.998081e-04  2.729782e-04  1.387897e-04
#>  [4,] -0.0004918683 -9.237708e-05 -7.575250e-04 -7.329078e-04 -4.108248e-04
#>  [5,]  0.0006499262  1.079796e-04  1.023582e-03  1.005907e-03  6.365167e-04
#>  [6,] -0.0009572435 -1.660669e-04 -1.531043e-03 -1.479122e-03 -8.970645e-04
#>  [7,]  0.0004919983  8.440806e-05  7.901386e-04  7.871512e-04  5.130353e-04
#>  [8,] -0.0008282347 -1.411517e-04 -1.300682e-03 -1.247882e-03 -7.904871e-04
#>  [9,] -0.0006374046 -1.116345e-04 -1.032153e-03 -9.704326e-04 -5.676827e-04
#> [10,] -0.0002598164 -4.423042e-05 -3.954914e-04 -3.609603e-04 -1.738127e-04
#>                 V23           V24           V25           V26           V27
#>  [1,]  1.624207e-05  6.660064e-07 -9.779734e-06 -1.545315e-05 -3.518363e-05
#>  [2,]  6.122331e-04  6.075649e-04  5.229992e-04  7.436131e-04  1.130316e-03
#>  [3,]  6.018000e-05  1.005657e-04  1.474632e-04  2.019045e-04  3.047359e-04
#>  [4,] -2.694294e-04 -3.087423e-04 -3.604658e-04 -4.923910e-04 -6.571882e-04
#>  [5,]  4.844792e-04  5.091178e-04  4.887247e-04  6.508593e-04  9.423517e-04
#>  [6,] -6.742650e-04 -7.112090e-04 -6.405058e-04 -8.672484e-04 -1.333653e-03
#>  [7,]  4.058323e-04  4.128836e-04  3.859058e-04  5.150499e-04  7.260455e-04
#>  [8,] -6.316545e-04 -6.379091e-04 -5.254425e-04 -7.003247e-04 -1.131607e-03
#>  [9,] -4.226614e-04 -4.236720e-04 -3.276127e-04 -4.696634e-04 -8.463333e-04
#> [10,] -8.527992e-05 -1.262181e-04 -1.560258e-04 -2.198910e-04 -3.417160e-04
#>                 V28           V29            V3           V30           V31
#>  [1,] -5.661794e-05 -0.0000671662 -3.936227e-06 -6.930979e-05 -0.0000407519
#>  [2,]  1.395320e-03  0.0012508929  1.223588e-04  1.023770e-03  0.0004869957
#>  [3,]  3.653918e-04  0.0003251772  2.158400e-05  2.749509e-04  0.0001357005
#>  [4,] -8.106745e-04 -0.0007982141 -8.090367e-05 -7.586573e-04 -0.0005088690
#>  [5,]  1.131970e-03  0.0010099541  8.431472e-05  8.194422e-04  0.0003988244
#>  [6,] -1.642088e-03 -0.0014668072 -1.280335e-04 -1.180826e-03 -0.0005106786
#>  [7,]  8.722305e-04  0.0007951042  6.749584e-05  6.606346e-04  0.0003460511
#>  [8,] -1.385110e-03 -0.0011887940 -1.092045e-04 -9.097880e-04 -0.0003233944
#>  [9,] -1.055116e-03 -0.0008758821 -8.535818e-05 -6.237767e-04 -0.0001128359
#> [10,] -4.437369e-04 -0.0004363313 -3.176645e-05 -3.951287e-04 -0.0002132147
#>                 V32           V33           V34           V35           V36
#>  [1,] -2.655411e-05 -1.018872e-05  9.616905e-06  2.227236e-05  5.505843e-05
#>  [2,]  4.841960e-04  1.146901e-04 -2.271241e-04 -5.293949e-04 -1.182339e-03
#>  [3,]  9.834682e-05 -1.179161e-05 -1.540920e-04 -2.298964e-04 -3.587259e-04
#>  [4,] -4.626972e-04 -2.788814e-04 -1.743408e-04 -6.756537e-05  3.173429e-04
#>  [5,]  3.620164e-04  7.092342e-05 -2.302666e-04 -4.673694e-04 -9.257115e-04
#>  [6,] -4.605279e-04 -2.839818e-05  4.476254e-04  8.202491e-04  1.529601e-03
#>  [7,]  3.311650e-04  1.245276e-04 -6.416599e-05 -2.274602e-04 -5.883385e-04
#>  [8,] -2.885808e-04  1.077082e-04  5.944259e-04  9.504058e-04  1.537850e-03
#>  [9,] -9.102207e-05  2.460065e-04  6.849001e-04  1.007676e-03  1.490481e-03
#> [10,] -1.773203e-04 -5.915650e-05  6.055251e-05  1.494747e-04  3.575792e-04
#>                 V37           V38           V39            V4           V40
#>  [1,]  3.457878e-05 -3.500984e-07 -1.474237e-05 -8.221689e-06 -3.473971e-06
#>  [2,] -9.201861e-04 -1.078100e-04  2.972218e-04  2.630178e-04  2.261844e-05
#>  [3,] -2.521979e-04 -4.881274e-05  7.638261e-05  5.080774e-05  2.348385e-05
#>  [4,]  2.568445e-04 -1.009601e-04 -2.521655e-04 -1.501069e-04 -8.750340e-05
#>  [5,] -6.954415e-04 -8.722274e-05  2.390382e-04  1.851910e-04  4.836485e-05
#>  [6,]  1.145837e-03  1.822250e-04 -3.220634e-04 -2.823784e-04 -2.858139e-05
#>  [7,] -4.595520e-04 -1.696756e-05  2.077132e-04  1.447930e-04  5.004470e-05
#>  [8,]  1.138128e-03  2.672777e-04 -2.044752e-04 -2.438484e-04  1.899652e-05
#>  [9,]  1.101603e-03  3.414997e-04 -9.307678e-05 -1.981058e-04  8.777375e-05
#> [10,]  2.686866e-04  1.707421e-05 -1.087161e-04 -7.070453e-05 -2.170535e-05
#>                 V41           V42           V43           V44           V45
#>  [1,] -2.333791e-05 -0.0000548627 -5.282691e-05 -4.815949e-05 -6.720612e-05
#>  [2,]  3.714799e-04  0.0009194105  1.027766e-03  9.043309e-04  1.202628e-03
#>  [3,]  8.390016e-05  0.0002118386  2.200566e-04  1.975877e-04  2.675673e-04
#>  [4,] -2.399207e-04 -0.0004937410 -5.460317e-04 -4.814359e-04 -6.554706e-04
#>  [5,]  2.829938e-04  0.0006834354  7.533666e-04  6.659336e-04  8.708647e-04
#>  [6,] -4.094862e-04 -0.0010613585 -1.169281e-03 -1.029280e-03 -1.352347e-03
#>  [7,]  2.239086e-04  0.0005204760  5.853648e-04  5.127264e-04  6.637830e-04
#>  [8,] -3.281031e-04 -0.0009060419 -9.960482e-04 -8.813290e-04 -1.158908e-03
#>  [9,] -1.819521e-04 -0.0006773465 -7.708634e-04 -6.826447e-04 -9.145814e-04
#> [10,] -1.244156e-04 -0.0003022540 -3.235718e-04 -2.881545e-04 -3.859909e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -6.436122e-05 -4.083214e-05 -2.735003e-05 -1.697798e-05 -1.115658e-05
#>  [2,]  1.119959e-03  7.077923e-04  4.870692e-04  3.106501e-04  2.602415e-04
#>  [3,]  2.625602e-04  1.820381e-04  1.236769e-04  7.775885e-05  6.945937e-05
#>  [4,] -6.339580e-04 -3.951318e-04 -2.709614e-04 -1.607551e-04 -1.377432e-04
#>  [5,]  8.175683e-04  5.330981e-04  3.632021e-04  2.318310e-04  2.028573e-04
#>  [6,] -1.264399e-03 -8.228012e-04 -5.600651e-04 -3.616353e-04 -3.055295e-04
#>  [7,]  6.182852e-04  3.925865e-04  2.687213e-04  1.716230e-04  1.495579e-04
#>  [8,] -1.084798e-03 -7.271087e-04 -4.919595e-04 -3.202871e-04 -2.713789e-04
#>  [9,] -8.716924e-04 -6.006115e-04 -4.017817e-04 -2.641583e-04 -2.252774e-04
#> [10,] -3.651713e-04 -2.317901e-04 -1.571740e-04 -9.929154e-05 -8.112736e-05
#>                 V50           V51           V52           V53           V54
#>  [1,] -4.806210e-06 -5.110778e-06 -3.706045e-06 -2.226869e-06 -1.216181e-06
#>  [2,]  8.564021e-05  9.443658e-05  7.687884e-05  4.083314e-05  3.228143e-05
#>  [3,]  2.042221e-05  2.215675e-05  1.620694e-05  9.485617e-06  6.426573e-06
#>  [4,] -5.031016e-05 -5.671349e-05 -4.736868e-05 -2.547255e-05 -2.075964e-05
#>  [5,]  6.365507e-05  6.897878e-05  5.467461e-05  3.043065e-05  2.308118e-05
#>  [6,] -9.691206e-05 -1.059814e-04 -8.379120e-05 -4.590175e-05 -3.434445e-05
#>  [7,]  4.876946e-05  5.268258e-05  4.246070e-05  2.371046e-05  1.878712e-05
#>  [8,] -8.124954e-05 -9.062773e-05 -7.090134e-05 -3.692830e-05 -2.700894e-05
#>  [9,] -6.442696e-05 -7.239556e-05 -5.556165e-05 -2.781956e-05 -1.936118e-05
#> [10,] -2.828947e-05 -3.035494e-05 -2.356221e-05 -1.378335e-05 -9.740251e-06
#>                 V55           V56           V57           V58           V59
#>  [1,] -5.331476e-07 -1.371472e-06 -1.268786e-07 -1.608908e-06  3.871614e-08
#>  [2,]  1.346206e-05  2.708919e-05  7.496342e-07  3.467722e-05  1.235404e-05
#>  [3,]  2.391490e-06  6.080778e-06  4.545195e-07  7.188036e-06  8.093504e-07
#>  [4,] -1.148259e-05 -1.758712e-05 -4.803922e-06 -2.112566e-05 -8.039428e-06
#>  [5,]  9.391419e-06  1.999977e-05  9.139684e-07  2.471467e-05  8.096444e-06
#>  [6,] -1.283673e-05 -2.989586e-05  1.443835e-07 -3.818632e-05 -1.163383e-05
#>  [7,]  8.193506e-06  1.570684e-05  1.346161e-06  1.959406e-05  7.370218e-06
#>  [8,] -7.739069e-06 -2.391765e-05  2.123720e-06 -3.153927e-05 -8.707348e-06
#>  [9,] -3.370729e-06 -1.815219e-05  3.957334e-06 -2.503346e-05 -5.585608e-06
#> [10,] -4.317426e-06 -8.889411e-06 -7.846533e-07 -1.061171e-05 -2.232669e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -1.181617e-05 -3.532968e-07 -1.791577e-05 -2.259869e-05 -3.877128e-05
#>  [2,]  2.166027e-04  1.371680e-05  2.913287e-04  4.490182e-04  8.508910e-04
#>  [3,]  6.432333e-05  2.555657e-06  9.002926e-05  1.049011e-04  1.857698e-04
#>  [4,] -1.276301e-04 -1.038344e-05 -1.757730e-04 -2.540971e-04 -4.578161e-04
#>  [5,]  1.783936e-04  9.654467e-06  2.398544e-04  3.366976e-04  6.199162e-04
#>  [6,] -2.648425e-04 -1.359188e-05 -3.609026e-04 -5.188601e-04 -9.651662e-04
#>  [7,]  1.298411e-04  8.278923e-06  1.755608e-04  2.536507e-04  4.684580e-04
#>  [8,] -2.344577e-04 -1.020737e-05 -3.153197e-04 -4.548473e-04 -8.481341e-04
#>  [9,] -1.904088e-04 -7.320882e-06 -2.606051e-04 -3.687238e-04 -6.908800e-04
#> [10,] -7.327035e-05 -3.810848e-06 -1.051495e-04 -1.395543e-04 -2.527187e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,] -0.01642297 -0.02140693 -0.005462757 -0.03632517 -0.01275277 -0.01138343
#> [2,]  0.01642297  0.02140693  0.005462757  0.03632517  0.01275277  0.01138343
#>             [,7]         [,8]         [,9]       [,10]
#> [1,] -0.01753151 -0.006429336 -0.004872591 -0.02193203
#> [2,]  0.01753151  0.006429336  0.004872591  0.02193203
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3998507 0.4499419 0.1580192 0.9194863 0.3493052 0.3487349 0.3609734
#>  [8] 0.2374562 0.1803987 0.5211954
#> 
#> $B[[2]]
#> [1]  0.02497777 -0.12937017
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -4.285156e-05  1.025193e-03  2.538387e-04 -7.945876e-04  8.361495e-04
#>  [6] -1.136158e-03  6.859097e-04 -8.539366e-04 -5.015047e-04 -3.378137e-04
#> 
#> $vB[[2]]
#> [1] -0.04189588  0.04189588
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#>  [2,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#>  [3,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [4,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#>  [5,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#>  [6,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#>  [7,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#>  [8,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#>  [9,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [10,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [11,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [12,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [13,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [14,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [15,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [16,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [17,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [18,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [19,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [20,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [21,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [22,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [23,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [24,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [25,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [26,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [27,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [28,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [29,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [30,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [31,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [32,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [33,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [34,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [35,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [36,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [37,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [38,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [39,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [40,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#>  [2,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#>  [3,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [4,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#>  [5,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#>  [6,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#>  [7,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#>  [8,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#>  [9,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [10,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [11,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [12,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [13,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [14,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [15,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [16,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [17,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [18,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [19,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [20,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [21,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [22,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [23,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [24,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [25,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [26,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [27,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [28,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [29,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [30,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [31,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [32,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [33,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [34,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [35,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [36,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [37,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [38,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [39,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [40,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#>  [2,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#>  [3,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [4,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#>  [5,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#>  [6,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#>  [7,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#>  [8,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#>  [9,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [10,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [11,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [12,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [13,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [14,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [15,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [16,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [17,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [18,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [19,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [20,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [21,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [22,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [23,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [24,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [25,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [26,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [27,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [28,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [29,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [30,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [31,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [32,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [33,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [34,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [35,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [36,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [37,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [38,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [39,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [40,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#>  [2,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#>  [3,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [4,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#>  [5,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#>  [6,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#>  [7,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#>  [8,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#>  [9,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [10,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [11,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [12,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [13,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [14,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [15,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [16,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [17,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [18,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [19,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [20,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [21,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [22,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [23,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [24,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [25,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [26,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [27,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [28,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [29,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [30,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [31,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [32,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [33,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [34,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [35,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [36,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [37,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [38,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [39,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [40,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#>  [2,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#>  [3,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [4,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#>  [5,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#>  [6,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#>  [7,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#>  [8,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#>  [9,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [10,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [11,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [12,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [13,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [14,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [15,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [16,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [17,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [18,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [19,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [20,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [21,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [22,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [23,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [24,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [25,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [26,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [27,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [28,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [29,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [30,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [31,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [32,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [33,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [34,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [35,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [36,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [37,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [38,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [39,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [40,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#>  [2,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#>  [3,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [4,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#>  [5,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#>  [6,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#>  [7,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#>  [8,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#>  [9,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [10,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [11,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [12,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [13,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [14,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [15,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [16,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [17,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [18,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [19,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [20,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [21,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [22,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [23,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [24,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [25,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [26,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [27,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [28,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [29,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [30,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [31,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [32,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [33,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [34,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [35,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [36,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [37,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [38,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [39,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [40,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6944039 0.6175398 0.3014001 0.9087909 0.5112444 0.5034160 0.5573305
#>  [2,] 0.6984102 0.7047658 0.4365503 0.9380557 0.6124227 0.5958994 0.7158004
#>  [3,] 0.7812008 0.7491571 0.3941787 0.9612315 0.6292132 0.5864791 0.7091973
#>  [4,] 0.5655492 0.5946818 0.2729709 0.9203574 0.4880749 0.4455910 0.6243473
#>  [5,] 0.5511838 0.5667315 0.3476584 0.8877736 0.5058583 0.4936044 0.5964674
#>  [6,] 0.7648044 0.6880882 0.3680802 0.9382408 0.6062457 0.5448365 0.6331409
#>  [7,] 0.7250947 0.5909966 0.3068902 0.9048502 0.5785254 0.4753033 0.5589049
#>  [8,] 0.6774254 0.5808296 0.3563696 0.9054797 0.5318332 0.4851610 0.5678035
#>  [9,] 0.6997143 0.6839670 0.4356189 0.9189429 0.5971411 0.5704687 0.6864239
#> [10,] 0.7261114 0.6852916 0.3686348 0.9625626 0.5874770 0.5786799 0.7161212
#> [11,] 0.4782162 0.4740687 0.2463415 0.8452322 0.3975452 0.3710609 0.4904274
#> [12,] 0.7174546 0.6988694 0.4275004 0.9383729 0.6126336 0.6129643 0.6916640
#> [13,] 0.7225281 0.7358133 0.4534640 0.9459132 0.6198674 0.6217556 0.7382898
#> [14,] 0.6535863 0.6508472 0.3916436 0.9071747 0.5318674 0.5069918 0.6096417
#> [15,] 0.5787174 0.5901398 0.3293048 0.9242172 0.4795210 0.5244392 0.6218388
#> [16,] 0.7478990 0.7362756 0.4227614 0.9598304 0.6404879 0.6341919 0.7427234
#> [17,] 0.5359765 0.4621148 0.2227436 0.8782019 0.4057141 0.3915281 0.4427796
#> [18,] 0.5730914 0.5798460 0.2887741 0.9059611 0.4527192 0.4106707 0.5634293
#> [19,] 0.5807232 0.5573428 0.3529300 0.8814183 0.5145660 0.5163104 0.6091935
#> [20,] 0.7297889 0.7217247 0.4088262 0.9608607 0.6608725 0.6423314 0.7692892
#> [21,] 0.7622093 0.7678791 0.4364763 0.9579203 0.6587501 0.6242155 0.7493458
#> [22,] 0.7101668 0.6979904 0.3905497 0.9250047 0.5907465 0.5911480 0.6647784
#> [23,] 0.7974902 0.7834810 0.4585686 0.9696556 0.6827050 0.6681448 0.7881663
#> [24,] 0.7106199 0.7074623 0.3767543 0.9440575 0.6078290 0.5996198 0.6995855
#> [25,] 0.7009983 0.6884159 0.3941555 0.9353700 0.6072244 0.5992984 0.7174683
#> [26,] 0.7971022 0.7752040 0.4358678 0.9721666 0.6842759 0.6581886 0.7903799
#> [27,] 0.7833724 0.7209609 0.3815249 0.9360474 0.6293009 0.5691682 0.6641873
#> [28,] 0.6446209 0.5309272 0.3397544 0.8668891 0.5073314 0.4664670 0.5206777
#> [29,] 0.7058660 0.7199799 0.4594590 0.9390832 0.6178513 0.6089499 0.7245875
#> [30,] 0.6717948 0.6362073 0.3314597 0.9249438 0.5172449 0.5129308 0.5929929
#> [31,] 0.7043520 0.7167427 0.4581448 0.9432970 0.6199204 0.6174758 0.7297129
#> [32,] 0.6479835 0.6338603 0.3116679 0.9436068 0.5051352 0.4510116 0.6232229
#> [33,] 0.7157777 0.6795312 0.3358250 0.9378394 0.5296485 0.5220721 0.6493275
#> [34,] 0.5906309 0.5837359 0.3389491 0.9265570 0.4823262 0.5198112 0.6066155
#> [35,] 0.7542233 0.7569696 0.4669362 0.9565247 0.6493913 0.6387902 0.7617971
#> [36,] 0.8200152 0.7623562 0.4061968 0.9630404 0.6862936 0.5976763 0.7246892
#> [37,] 0.7228211 0.7088854 0.4575992 0.9499122 0.6341437 0.6193301 0.7391229
#> [38,] 0.7766421 0.7496292 0.3658255 0.9637411 0.6224940 0.5801218 0.7059932
#> [39,] 0.7987389 0.7720437 0.3908581 0.9678571 0.6503885 0.6002279 0.7331763
#> [40,] 0.6026681 0.6311651 0.3422640 0.9360569 0.5072725 0.5476240 0.6587636
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3615603 0.2884228 0.5590446
#>  [2,] 0.5153509 0.3714797 0.7158722
#>  [3,] 0.4760164 0.3470806 0.7185788
#>  [4,] 0.3651890 0.2220020 0.6085765
#>  [5,] 0.4102897 0.2972789 0.6141517
#>  [6,] 0.4482207 0.3355352 0.6451673
#>  [7,] 0.3830484 0.2697709 0.5654497
#>  [8,] 0.3133175 0.2594154 0.5459818
#>  [9,] 0.4805723 0.3771254 0.6713209
#> [10,] 0.4774203 0.3061491 0.6785624
#> [11,] 0.2988364 0.2051312 0.4992057
#> [12,] 0.5048228 0.3868727 0.6794740
#> [13,] 0.5429073 0.3998806 0.7527873
#> [14,] 0.3912873 0.3205787 0.6120610
#> [15,] 0.4105974 0.2670667 0.5986199
#> [16,] 0.5529329 0.3781220 0.7348465
#> [17,] 0.2609847 0.1848997 0.4472354
#> [18,] 0.3257700 0.2422258 0.5758025
#> [19,] 0.4312817 0.3074564 0.5629150
#> [20,] 0.5941335 0.3500768 0.7611735
#> [21,] 0.5564991 0.4056875 0.7662023
#> [22,] 0.4737930 0.3817290 0.6621043
#> [23,] 0.5982225 0.4057724 0.7822245
#> [24,] 0.5133064 0.3505544 0.7095890
#> [25,] 0.5222595 0.3562552 0.6990324
#> [26,] 0.5878809 0.3830577 0.7770613
#> [27,] 0.5010541 0.3750518 0.6870865
#> [28,] 0.3351301 0.2563751 0.5063524
#> [29,] 0.5301158 0.3982477 0.7307914
#> [30,] 0.3669248 0.2880486 0.5857216
#> [31,] 0.5374590 0.3946443 0.7386765
#> [32,] 0.3572037 0.2415318 0.6246912
#> [33,] 0.3838189 0.3102453 0.6159055
#> [34,] 0.3991700 0.2662386 0.5917811
#> [35,] 0.5594942 0.4023755 0.7732272
#> [36,] 0.5132658 0.3643340 0.7394579
#> [37,] 0.5460033 0.3622164 0.7460084
#> [38,] 0.4698063 0.3366039 0.7177566
#> [39,] 0.4874997 0.3527442 0.7352795
#> [40,] 0.4343180 0.2862340 0.6498874
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5560135 0.4439865
#>  [2,] 0.5594625 0.4405375
#>  [3,] 0.5578936 0.4421064
#>  [4,] 0.5540423 0.4459577
#>  [5,] 0.5569325 0.4430675
#>  [6,] 0.5573668 0.4426332
#>  [7,] 0.5546544 0.4453456
#>  [8,] 0.5533551 0.4466449
#>  [9,] 0.5580925 0.4419075
#> [10,] 0.5587807 0.4412193
#> [11,] 0.5524577 0.4475423
#> [12,] 0.5602794 0.4397206
#> [13,] 0.5608195 0.4391805
#> [14,] 0.5552774 0.4447226
#> [15,] 0.5581711 0.4418289
#> [16,] 0.5611457 0.4388543
#> [17,] 0.5536082 0.4463918
#> [18,] 0.5534649 0.4465351
#> [19,] 0.5574369 0.4425631
#> [20,] 0.5616740 0.4383260
#> [21,] 0.5605407 0.4394593
#> [22,] 0.5592155 0.4407845
#> [23,] 0.5618526 0.4381474
#> [24,] 0.5599690 0.4400310
#> [25,] 0.5599160 0.4400840
#> [26,] 0.5612040 0.4387960
#> [27,] 0.5587507 0.4412493
#> [28,] 0.5537242 0.4462758
#> [29,] 0.5601924 0.4398076
#> [30,] 0.5559669 0.4440331
#> [31,] 0.5607320 0.4392680
#> [32,] 0.5541237 0.4458763
#> [33,] 0.5559191 0.4440809
#> [34,] 0.5580448 0.4419552
#> [35,] 0.5609205 0.4390795
#> [36,] 0.5581435 0.4418565
#> [37,] 0.5604075 0.4395925
#> [38,] 0.5578117 0.4421883
#> [39,] 0.5578815 0.4421185
#> [40,] 0.5586816 0.4413184
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]       [,2]       [,3]     [,4]        [,5]        [,6]
#>  [1,]  0.82078950  0.4791189 -0.8406396 2.298960  0.04498503  0.01366434
#>  [2,]  0.83973872  0.8700965 -0.2551743 2.717574  0.45750762  0.38840804
#>  [3,]  1.27267786  1.0941219 -0.4297809 3.210607  0.52884310  0.34942909
#>  [4,]  0.26371462  0.3833538 -0.9796011 2.447213 -0.04770927 -0.21850107
#>  [5,]  0.20545471  0.2685282 -0.6293478 2.068198  0.02343441 -0.02558370
#>  [6,]  1.17920254  0.7911965 -0.5404615 2.720764  0.43155830  0.17982909
#>  [7,]  0.96987541  0.3680868 -0.8146983 2.252317  0.31672314 -0.09886739
#>  [8,]  0.74196503  0.3261800 -0.5911564 2.259650  0.12750533 -0.05937338
#>  [9,]  0.84593758  0.7720630 -0.2589618 2.428070  0.39356713  0.28376372
#> [10,]  0.97498205  0.7781981 -0.5380782 3.246930  0.35354515  0.31735657
#> [11,] -0.08719053 -0.1038184 -1.1182205 1.697685 -0.41570419 -0.52766815
#> [12,]  0.93187056  0.8419197 -0.2920570 2.723047  0.45839612  0.45978963
#> [13,]  0.95703643  1.0243203 -0.1866842 2.861561  0.48898555  0.49700636
#> [14,]  0.63484066  0.6227654 -0.4404085 2.279616  0.12764267  0.02796898
#> [15,]  0.31751051  0.3645435 -0.7113311 2.501076 -0.08196200  0.09783488
#> [16,]  1.08743827  1.0266997 -0.3114478 3.173646  0.57748245  0.55024278
#> [17,]  0.14415506 -0.1518319 -1.2497490 1.975512 -0.38171174 -0.44089365
#> [18,]  0.29447528  0.3221414 -0.9013452 2.265288 -0.18968994 -0.36119339
#> [19,]  0.32574272  0.2303849 -0.6061848 2.005930  0.05828058  0.06526495
#> [20,]  0.99355174  0.9530326 -0.3688199 3.200703  0.66718460  0.58549734
#> [21,]  1.16483055  1.1963742 -0.2554752 3.125199  0.65772905  0.50747963
#> [22,]  0.89619452  0.8377465 -0.4450023 2.512373  0.36705231  0.36871319
#> [23,]  1.37068152  1.2860687 -0.1661066 3.464330  0.76623087  0.69980636
#> [24,]  0.89839650  0.8830908 -0.5033476 2.825862  0.43819566  0.40388130
#> [25,]  0.85205612  0.7927241 -0.4298777 2.672264  0.43565984  0.40254272
#> [26,]  1.36828074  1.2379327 -0.2579496 3.553292  0.77349256  0.65523241
#> [27,]  1.28542844  0.9492328 -0.4830810 2.683524  0.52921883  0.27845838
#> [28,]  0.59547738  0.1238668 -0.6643889 1.873728  0.02932769 -0.13433345
#> [29,]  0.87538992  0.9443621 -0.1625208 2.735396  0.48043804  0.44290018
#> [30,]  0.71631377  0.5589404 -0.7015906 2.511495  0.06900714  0.05173479
#> [31,]  0.86810886  0.9283610 -0.1678137 2.811555  0.48921059  0.47884786
#> [32,]  0.61018725  0.5488135 -0.7923332 2.817360  0.02054139 -0.19658428
#> [33,]  0.92361278  0.7516180 -0.6819554 2.713858  0.11873327  0.08834590
#> [34,]  0.36657406  0.3381288 -0.6679810 2.534966 -0.07072480  0.07928649
#> [35,]  1.12126513  1.1361366 -0.1324483 3.091114  0.61636469  0.57011726
#> [36,]  1.51645032  1.1656412 -0.3797103 3.260269  0.78284790  0.39579223
#> [37,]  0.95849850  0.8899766 -0.1700114 2.942592  0.55003515  0.48670583
#> [38,]  1.24620426  1.0966358 -0.5501678 3.280138  0.50014755  0.32327340
#> [39,]  1.37843126  1.2198870 -0.4437066 3.404892  0.62074718  0.40641496
#> [40,]  0.41659471  0.5372184 -0.6532211 2.683682  0.02909195  0.19107523
#>              [,7]         [,8]       [,9]        [,10]
#>  [1,]  0.23033506 -0.568598209 -0.9030566  0.237285683
#>  [2,]  0.92372445  0.061423070 -0.5258742  0.924077561
#>  [3,]  0.89148850 -0.096008190 -0.6318968  0.937422996
#>  [4,]  0.50804159 -0.552912221 -1.2540377  0.441332626
#>  [5,]  0.39076747 -0.362767939 -0.8602895  0.464797752
#>  [6,]  0.54571508 -0.207862561 -0.6832552  0.597863036
#>  [7,]  0.23671876 -0.476629105 -0.9957851  0.263309563
#>  [8,]  0.27289505 -0.784655078 -1.0490094  0.184448385
#>  [9,]  0.78345333 -0.077749996 -0.5017675  0.714165453
#> [10,]  0.92530223 -0.090380139 -0.8181847  0.747173149
#> [11,] -0.03829507 -0.852844894 -1.3545274 -0.003177398
#> [12,]  0.80791011  0.019291734 -0.4604767  0.751355377
#> [13,]  1.03709862  0.172052273 -0.4059626  1.113533421
#> [14,]  0.44580648 -0.441904534 -0.7511136  0.455983721
#> [15,]  0.49736036 -0.361496308 -1.0095560  0.399718126
#> [16,]  1.06017192  0.212527825 -0.4975271  1.019352682
#> [17,] -0.22988867 -1.040857059 -1.4834974 -0.211847238
#> [18,]  0.25509140 -0.727379785 -1.1405152  0.305565694
#> [19,]  0.44392332 -0.276623855 -0.8120379  0.253000786
#> [20,]  1.20430208  0.381079687 -0.6187017  1.159123753
#> [21,]  1.09512600  0.226965616 -0.3818221  1.186990165
#> [22,]  0.68466196 -0.104924323 -0.4822161  0.672685646
#> [23,]  1.31390746  0.398064135 -0.3814698  1.278677345
#> [24,]  0.84532502  0.053238255 -0.6166030  0.893388762
#> [25,]  0.93193803  0.089096762 -0.5916549  0.842694551
#> [26,]  1.32721691  0.355211890 -0.4765898  1.248622395
#> [27,]  0.68201079  0.004216371 -0.5106048  0.786533554
#> [28,]  0.08275798 -0.685072698 -1.0648953  0.025410957
#> [29,]  0.96733239  0.120609304 -0.4127719  0.998641519
#> [30,]  0.37635187 -0.545432301 -0.9048803  0.346306558
#> [31,]  0.99316650  0.150117101 -0.4278314  1.039100854
#> [32,]  0.50325045 -0.587521525 -1.1442997  0.509508130
#> [33,]  0.61608442 -0.473370275 -0.7989729  0.472205008
#> [34,]  0.43310764 -0.408924813 -1.0137911  0.371333216
#> [35,]  1.16255750  0.239109692 -0.3955769  1.226624537
#> [36,]  0.96784213  0.053075692 -0.5566021  1.043153104
#> [37,]  1.04141460  0.184534927 -0.5657571  1.077435787
#> [38,]  0.87600275 -0.120921860 -0.6784653  0.933360633
#> [39,]  1.01079784 -0.050011680 -0.6069985  1.021576142
#> [40,]  0.65778917 -0.264255149 -0.9137453  0.618544119
#> 
#> $pre[[3]]
#>              [,1]       [,2]
#>  [1,] 0.024287579 -0.2007108
#>  [2,] 0.010607254 -0.2283738
#>  [3,] 0.012665902 -0.2199517
#>  [4,] 0.009511121 -0.2075060
#>  [5,] 0.013631109 -0.2150908
#>  [6,] 0.018510462 -0.2119716
#>  [7,] 0.019514731 -0.1999798
#>  [8,] 0.016473751 -0.1977624
#>  [9,] 0.012318120 -0.2211060
#> [10,] 0.013405391 -0.2228095
#> [11,] 0.016171613 -0.1944342
#> [12,] 0.016512441 -0.2257837
#> [13,] 0.011377444 -0.2331112
#> [14,] 0.014044766 -0.2079723
#> [15,] 0.017182915 -0.2165601
#> [16,] 0.013216723 -0.2325962
#> [17,] 0.025944438 -0.1893156
#> [18,] 0.014471733 -0.2002086
#> [19,] 0.016612049 -0.2141544
#> [20,] 0.009202664 -0.2387559
#> [21,] 0.011191207 -0.2321656
#> [22,] 0.018277861 -0.2197009
#> [23,] 0.010609353 -0.2380748
#> [24,] 0.014466213 -0.2265701
#> [25,] 0.012802405 -0.2280188
#> [26,] 0.009625681 -0.2364240
#> [27,] 0.017951997 -0.2181415
#> [28,] 0.019442346 -0.1962874
#> [29,] 0.011038523 -0.2309045
#> [30,] 0.019673216 -0.2051364
#> [31,] 0.011151752 -0.2329817
#> [32,] 0.011904347 -0.2054420
#> [33,] 0.017695659 -0.2069202
#> [34,] 0.018237933 -0.2149927
#> [35,] 0.009733254 -0.2351653
#> [36,] 0.011188120 -0.2224430
#> [37,] 0.008972029 -0.2338441
#> [38,] 0.013379527 -0.2189063
#> [39,] 0.011310055 -0.2212587
#> [40,] 0.014690132 -0.2211229
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5560135  0.5560135
#>  [2,]  0.4405375 -0.4405375
#>  [3,]  0.4421064 -0.4421064
#>  [4,]  0.4459577 -0.4459577
#>  [5,]  0.4430675 -0.4430675
#>  [6,]  0.4426332 -0.4426332
#>  [7,]  0.4453456 -0.4453456
#>  [8,] -0.5533551  0.5533551
#>  [9,] -0.5580925  0.5580925
#> [10,] -0.5587807  0.5587807
#> [11,]  0.4475423 -0.4475423
#> [12,] -0.5602794  0.5602794
#> [13,]  0.4391805 -0.4391805
#> [14,]  0.4447226 -0.4447226
#> [15,]  0.4418289 -0.4418289
#> [16,] -0.5611457  0.5611457
#> [17,]  0.4463918 -0.4463918
#> [18,]  0.4465351 -0.4465351
#> [19,] -0.5574369  0.5574369
#> [20,] -0.5616740  0.5616740
#> [21,]  0.4394593 -0.4394593
#> [22,]  0.4407845 -0.4407845
#> [23,] -0.5618526  0.5618526
#> [24,]  0.4400310 -0.4400310
#> [25,] -0.5599160  0.5599160
#> [26,] -0.5612040  0.5612040
#> [27,] -0.5587507  0.5587507
#> [28,]  0.4462758 -0.4462758
#> [29,]  0.4398076 -0.4398076
#> [30,] -0.5559669  0.5559669
#> [31,]  0.4392680 -0.4392680
#> [32,]  0.4458763 -0.4458763
#> [33,] -0.5559191  0.5559191
#> [34,]  0.4419552 -0.4419552
#> [35,]  0.4390795 -0.4390795
#> [36,] -0.5581435  0.5581435
#> [37,] -0.5604075  0.5604075
#> [38,]  0.4421883 -0.4421883
#> [39,]  0.4421185 -0.4421185
#> [40,]  0.4413184 -0.4413184
#> 
#> $L
#> [1] 0.6898948 0.6865821 0.6856467 0.6894464 0.6898046 0.6779657
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319