Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11           V12          V13
#>  [1,] -0.3601455 -0.105289355 -0.15332588  0.0040063198 -0.062510083
#>  [2,] -0.3776352 -0.112367085 -0.25994169 -0.1587964818 -0.239354710
#>  [3,] -0.3840169  0.009720607  0.02847081 -0.0656347460  0.005223559
#>  [4,] -0.2267785 -0.266335183 -0.12230740 -0.1444234447 -0.179851478
#>  [5,] -0.2804716 -0.114184342 -0.07105250 -0.1133976145  0.006386400
#>  [6,] -0.3352092 -0.187608897 -0.16926854 -0.1841294612 -0.119001906
#>  [7,] -0.3227165 -0.100726179 -0.18283756 -0.0005295442 -0.078193883
#>  [8,] -0.3263940 -0.199173293 -0.05593730 -0.0232302161 -0.116750174
#>  [9,] -0.3316180 -0.077072909 -0.08819281 -0.0856486527 -0.131884760
#> [10,] -0.2774342 -0.150481211 -0.22678336 -0.1733199258 -0.130939715
#>                V14           V15         V16          V17         V18
#>  [1,]  0.022402798  0.0607907559  0.11157374  0.002731433  0.09628557
#>  [2,] -0.116787240 -0.1459388580 -0.07360791 -0.176176951 -0.13694135
#>  [3,] -0.001154589 -0.0334405073 -0.05201137 -0.035320850 -0.02137242
#>  [4,] -0.020357176 -0.0984765705 -0.06165459  0.048586019 -0.05286005
#>  [5,]  0.043384372 -0.0717529161 -0.01185722  0.041325027  0.08377871
#>  [6,] -0.082029941 -0.0285800126 -0.13720204 -0.112346532  0.05911029
#>  [7,] -0.016161171 -0.0002738492 -0.12543094 -0.029926542 -0.13881592
#>  [8,] -0.148059068 -0.0637529938  0.02393219 -0.026057986  0.07390557
#>  [9,] -0.086793839 -0.0223764114 -0.08855689 -0.063301855 -0.10344561
#> [10,] -0.132233218 -0.1710429378 -0.12502297 -0.080566225 -0.16473685
#>               V19         V2         V20          V21         V22         V23
#>  [1,]  0.12607204 -0.3857849  0.12298023  0.092824735  0.01385898  0.04947617
#>  [2,] -0.04697714 -0.2405973 -0.03019576  0.107211175  0.10116557 -0.05620059
#>  [3,]  0.10516325 -0.3232987  0.08478543  0.239782651  0.23711506  0.20006787
#>  [4,] -0.11175112 -0.3168779  0.08063038  0.083946564 -0.03184726  0.04825205
#>  [5,]  0.11682070 -0.3440364  0.15525708  0.085514230  0.01934700  0.16588719
#>  [6,] -0.05659514 -0.2908712  0.01457518  0.084161283  0.07534200  0.15468565
#>  [7,] -0.04672681 -0.1573441  0.01308476  0.018795481 -0.02273962 -0.11305296
#>  [8,]  0.10256231 -0.3185849 -0.03207650  0.004666038  0.17798505  0.18353466
#>  [9,] -0.09559543 -0.2899476  0.09570480  0.103673710  0.18618294  0.20838372
#> [10,] -0.04937668 -0.1724657 -0.11482765 -0.067682246 -0.01095543  0.03196590
#>               V24         V25         V26         V27         V28         V29
#>  [1,]  0.21647937  0.03946903 0.080926390 0.128389470  0.11442365  0.08379049
#>  [2,]  0.04779489  0.07570539 0.035538777 0.049182981  0.01115140 -0.08613522
#>  [3,]  0.19601398  0.19062175 0.183868831 0.277024741  0.24975830  0.19665205
#>  [4,] -0.01658711  0.16792847 0.191826860 0.066558595 -0.02432366 -0.04434245
#>  [5,]  0.07416781  0.15610126 0.193492756 0.246196991  0.10841823  0.18855727
#>  [6,]  0.05259498  0.13517659 0.112434390 0.153945168  0.20120223  0.02544967
#>  [7,] -0.07632508 -0.04947880 0.007972289 0.003795796 -0.07914972 -0.04242509
#>  [8,]  0.30986111  0.32933427 0.218316174 0.144894450  0.20288788  0.28566044
#>  [9,]  0.13209032  0.09237787 0.110148191 0.128268241  0.14366737  0.08880836
#> [10,] -0.08938272 -0.02736641 0.048222419 0.008968169 -0.01259734  0.03949988
#>               V3         V30         V31         V32          V33         V34
#>  [1,] -0.2368821 -0.03952478 -0.01494344 -0.11613507 -0.128039266 -0.16239321
#>  [2,] -0.4233626  0.02057146  0.01566592 -0.01426975 -0.114150966 -0.11932857
#>  [3,] -0.2339066  0.20765326  0.01420503  0.12100918  0.140198137 -0.06024395
#>  [4,] -0.3696016 -0.13048584 -0.18082744 -0.02375734 -0.170483213 -0.22534846
#>  [5,] -0.2122818  0.07560043  0.11409022  0.13464994  0.055903994  0.12855797
#>  [6,] -0.1902972 -0.01266861 -0.03930378 -0.03579971 -0.123271297 -0.12200707
#>  [7,] -0.2364640 -0.07307914 -0.11319534  0.02508566 -0.124847323 -0.08002437
#>  [8,] -0.3396858  0.13078817  0.18689602  0.04240600 -0.029076295  0.11417634
#>  [9,] -0.3043731  0.08942782  0.06091177  0.04959504  0.017662836  0.03515889
#> [10,] -0.3388648 -0.07568555  0.05961884 -0.01951184  0.009556255 -0.01250016
#>               V35          V36          V37         V38           V39
#>  [1,] -0.05520083 -0.154222169 -0.094742552 -0.20250726 -0.0733645913
#>  [2,] -0.02160322 -0.004757565 -0.030592060 -0.13593773 -0.0436560433
#>  [3,]  0.14041543  0.118348356  0.065293626  0.05050862 -0.0005553762
#>  [4,] -0.17183553 -0.044448261 -0.098992162 -0.14688984 -0.2096895571
#>  [5,]  0.02448805  0.051273878  0.074991891  0.11544607  0.0090053210
#>  [6,] -0.05161973  0.007203847  0.005641405 -0.16592036 -0.0525126607
#>  [7,] -0.15167883 -0.092866639 -0.172139132 -0.05476917 -0.1728945118
#>  [8,] -0.02302872 -0.026640453  0.012395262  0.06184063  0.0654740580
#>  [9,] -0.05143822 -0.043792792 -0.082330113 -0.12579867 -0.0036888182
#> [10,]  0.01351015 -0.090827239  0.005987030  0.04617694 -0.1444806856
#>               V4          V40          V41          V42         V43         V44
#>  [1,] -0.2377285 -0.134813783 -0.104216018 -0.134040946 -0.23382760 -0.23154476
#>  [2,] -0.3664014 -0.069628242 -0.164720462 -0.180318818 -0.16420137 -0.09166754
#>  [3,] -0.3373239 -0.091904366 -0.054226579 -0.070218016 -0.06484830 -0.12425566
#>  [4,] -0.3424100 -0.057614580 -0.220028012 -0.210394875 -0.18695598 -0.12003987
#>  [5,] -0.2575902  0.004336484 -0.002116598  0.009466528 -0.01638468 -0.08960585
#>  [6,] -0.1861856 -0.090684987 -0.004562901 -0.128880699 -0.08292569 -0.11539195
#>  [7,] -0.2729017 -0.035648253 -0.049185083 -0.156401247 -0.22305342 -0.18611202
#>  [8,] -0.2297237 -0.021823468 -0.090093899 -0.052710391  0.01170903 -0.10586290
#>  [9,] -0.2849622 -0.042075005 -0.117340075 -0.018092591 -0.13023580 -0.11075593
#> [10,] -0.2866589 -0.079871599 -0.062408625  0.002652103 -0.11542574 -0.07814329
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.12671259 -0.13395607 -0.2647435 -0.3084319 -0.3367245 -0.2865449
#>  [2,] -0.12943625 -0.20020576 -0.2147293 -0.2723927 -0.2582042 -0.2896248
#>  [3,] -0.12559759 -0.16701158 -0.1732413 -0.2781819 -0.3257299 -0.1474338
#>  [4,] -0.12668327 -0.29480460 -0.1803177 -0.1690548 -0.2379498 -0.3555958
#>  [5,] -0.03949226 -0.10665201 -0.1430590 -0.2231246 -0.3422068 -0.1444082
#>  [6,] -0.15313340 -0.14561030 -0.2120306 -0.1760405 -0.2057623 -0.1281018
#>  [7,] -0.08175119 -0.20652679 -0.1205094 -0.2548995 -0.2936370 -0.2531407
#>  [8,] -0.19143040 -0.19248993 -0.1940986 -0.2494845 -0.3114526 -0.3504264
#>  [9,] -0.16581406 -0.07804167 -0.2055816 -0.0813363 -0.2273697 -0.1649718
#> [10,] -0.15920410 -0.07141759 -0.1744116 -0.2567870 -0.1869693 -0.1938382
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3863520 -0.3643477 -0.3101982 -0.2916603 -0.3175103 -0.2803443
#>  [2,] -0.4080427 -0.3492122 -0.4367594 -0.3472793 -0.2949676 -0.3601754
#>  [3,] -0.2407023 -0.4041963 -0.2832973 -0.3475614 -0.4062165 -0.3778099
#>  [4,] -0.3747039 -0.3174798 -0.2642619 -0.4256669 -0.2752236 -0.3083662
#>  [5,] -0.2645765 -0.2712540 -0.3251819 -0.3555265 -0.3959103 -0.3628398
#>  [6,] -0.2030792 -0.3920320 -0.4065438 -0.2318897 -0.3630250 -0.3545683
#>  [7,] -0.3818113 -0.2209906 -0.3830781 -0.2097485 -0.2652800 -0.3794949
#>  [8,] -0.3413077 -0.3735374 -0.2912716 -0.3687352 -0.3154504 -0.3727961
#>  [9,] -0.2929380 -0.2966605 -0.2928588 -0.3724901 -0.2725289 -0.2441385
#> [10,] -0.3350925 -0.3800625 -0.3244501 -0.2378131 -0.3878306 -0.3656188
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4175174 -0.3528861 -0.4258836 -0.2489931 -0.1958328 -0.3295154
#>  [2,] -0.3625993 -0.4763312 -0.4388299 -0.4474813 -0.2574369 -0.4017545
#>  [3,] -0.3489824 -0.3193986 -0.4151872 -0.4147423 -0.1093051 -0.3348806
#>  [4,] -0.2845987 -0.3098527 -0.2371315 -0.2407502 -0.1467165 -0.2780479
#>  [5,] -0.4208088 -0.3406498 -0.3115937 -0.2286099 -0.2722947 -0.3633126
#>  [6,] -0.2426752 -0.2896234 -0.2406250 -0.3663647 -0.2497062 -0.3427727
#>  [7,] -0.3075720 -0.4017390 -0.2740344 -0.2941074 -0.1446551 -0.3506967
#>  [8,] -0.3866362 -0.3236524 -0.3048772 -0.4155748 -0.2168607 -0.3268268
#>  [9,] -0.2502054 -0.2415684 -0.3535942 -0.3186319 -0.1026180 -0.2467141
#> [10,] -0.3329521 -0.2808488 -0.3365526 -0.3746081 -0.1953048 -0.3870724
#>                V7         V8          V9
#>  [1,] -0.21838899 -0.1276171 -0.09655629
#>  [2,] -0.16770735 -0.1957654 -0.15614127
#>  [3,] -0.15095982 -0.1864956 -0.02068064
#>  [4,] -0.18833127 -0.2866969 -0.14697654
#>  [5,] -0.05544018 -0.0623685 -0.07752704
#>  [6,] -0.25819493 -0.1967199 -0.13023581
#>  [7,] -0.24628445 -0.2680706 -0.10929066
#>  [8,] -0.20020866 -0.1193756 -0.23494860
#>  [9,] -0.22076549 -0.1410035 -0.03703655
#> [10,] -0.25273746 -0.1710100 -0.25232801
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.02890755 -0.09560482  0.01884452 -0.11761742  0.07883654 -0.09613268
#> [2,]  0.04870589  0.04094265 -0.09821134 -0.05166759 -0.01496126 -0.04347804
#>            [,7]        [,8]         [,9]       [,10]
#> [1,] -0.1045039  0.03831097 -0.003053889 -0.13032983
#> [2,]  0.0461498 -0.03149039  0.001529088  0.07895687
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1          V10           V11          V12          V13
#>  [1,]  1.696997e-04 0.0012322258  1.442690e-03 1.378233e-03 0.0012364677
#>  [2,]  1.393017e-04 0.0008952653  1.212166e-03 1.148353e-03 0.0008407118
#>  [3,] -5.656341e-05 0.0001151489 -5.859174e-05 8.669524e-05 0.0003500073
#>  [4,]  6.547567e-05 0.0004541631  5.874897e-04 5.951814e-04 0.0005082287
#>  [5,] -4.195063e-05 0.0001699018 -3.077712e-06 1.223133e-04 0.0003874735
#>  [6,]  1.130874e-04 0.0008980985  1.006341e-03 1.004968e-03 0.0009894784
#>  [7,]  1.284896e-04 0.0007559537  1.103213e-03 1.015778e-03 0.0006488972
#>  [8,] -4.390400e-05 0.0002137951  4.931720e-05 1.984600e-04 0.0004718939
#>  [9,]  8.023064e-05 0.0008433937  8.272401e-04 8.596843e-04 0.0009804647
#> [10,]  1.739042e-04 0.0010669556  1.529243e-03 1.379544e-03 0.0008417726
#>                V14           V15           V16           V17           V18
#>  [1,] 1.007937e-03  8.603528e-04  9.441360e-04  0.0010800127  1.320902e-03
#>  [2,] 3.595686e-04  8.253381e-05  3.610764e-05  0.0001684680  5.075815e-04
#>  [3,] 6.435483e-04  8.262575e-04  9.913105e-04  0.0010231491  9.248803e-04
#>  [4,] 3.535464e-04  2.833719e-04  3.202654e-04  0.0004084396  5.590648e-04
#>  [5,] 7.091706e-04  9.084694e-04  1.092998e-03  0.0011282572  1.023914e-03
#>  [6,] 9.463347e-04  9.285667e-04  1.066461e-03  0.0011794626  1.332271e-03
#>  [7,] 8.300665e-05 -2.529636e-04 -3.621291e-04 -0.0002166635  1.081049e-04
#>  [8,] 7.563247e-04  9.457631e-04  1.138950e-03  0.0012010021  1.108145e-03
#>  [9,] 1.137230e-03  1.241828e-03  1.448514e-03  0.0015482030  1.596514e-03
#> [10,] 2.988953e-05 -4.493951e-04 -6.191152e-04 -0.0004711989 -7.091363e-06
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0019263315  1.803715e-04 0.0024206076 0.0028119609 0.0026666564
#>  [2,] 0.0011620716  1.394930e-04 0.0016136043 0.0019176805 0.0016938753
#>  [3,] 0.0006700421 -5.288538e-05 0.0004578436 0.0004267871 0.0008367643
#>  [4,] 0.0008169438  6.871490e-05 0.0009809268 0.0011267881 0.0011081735
#>  [5,] 0.0007795220 -3.499914e-05 0.0006055411 0.0006068953 0.0009772617
#>  [6,] 0.0016660437  1.270474e-04 0.0019529682 0.0022083808 0.0022151156
#>  [7,] 0.0007285944  1.269665e-04 0.0011322364 0.0014247580 0.0011659734
#>  [8,] 0.0008676576 -3.294021e-05 0.0006535142 0.0006343881 0.0010590831
#>  [9,] 0.0017245346  9.885536e-05 0.0018842039 0.0020909139 0.0022409681
#> [10,] 0.0009252252  1.686541e-04 0.0015292213 0.0019127213 0.0014872226
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0024175443 0.0023473303 0.0021481037 2.193297e-03 2.148014e-03
#>  [2,] 0.0013163665 0.0011238408 0.0009027762 8.606535e-04 8.485471e-04
#>  [3,] 0.0012187502 0.0013689047 0.0014106444 1.493387e-03 1.560206e-03
#>  [4,] 0.0010110321 0.0009583639 0.0008624919 8.434784e-04 8.462966e-04
#>  [5,] 0.0013381629 0.0015108652 0.0015657034 1.620629e-03 1.676306e-03
#>  [6,] 0.0021590115 0.0021717249 0.0020633757 2.099383e-03 2.078292e-03
#>  [7,] 0.0007079499 0.0004387027 0.0001745689 4.532891e-05 3.708510e-05
#>  [8,] 0.0014495267 0.0015897486 0.0015920483 1.625624e-03 1.724361e-03
#>  [9,] 0.0023514365 0.0024403340 0.0023742532 2.420704e-03 2.408331e-03
#> [10,] 0.0008414621 0.0004929702 0.0001548734 2.355138e-05 2.538245e-05
#>                V28          V29            V3          V30           V31
#>  [1,] 0.0021744261 0.0019983088  2.146169e-04 1.821765e-03  0.0010471659
#>  [2,] 0.0010476450 0.0010047439  1.582730e-04 7.167413e-04 -0.0001518769
#>  [3,] 0.0012594514 0.0010843153 -5.250621e-05 1.116076e-03  0.0014282868
#>  [4,] 0.0008752004 0.0008021960  8.123953e-05 6.586524e-04  0.0003177991
#>  [5,] 0.0013826044 0.0011832548 -2.876052e-05 1.229725e-03  0.0015247256
#>  [6,] 0.0020243085 0.0018409076  1.514961e-04 1.711128e-03  0.0012740299
#>  [7,] 0.0002808119 0.0002949339  1.429820e-04 6.394808e-05 -0.0007539422
#>  [8,] 0.0014148003 0.0011874413 -2.231012e-05 1.231677e-03  0.0015628169
#>  [9,] 0.0022228521 0.0019826508  1.238707e-04 1.967281e-03  0.0017844264
#> [10,] 0.0004118029 0.0004276536  1.925514e-04 6.509641e-05 -0.0011535261
#>                 V32           V33           V34           V35           V36
#>  [1,]  1.012968e-03  8.865209e-04  0.0005505159  1.382754e-04 -3.918264e-05
#>  [2,]  1.510647e-05 -9.636835e-06 -0.0002848121 -6.324915e-04 -7.056498e-04
#>  [3,]  1.090789e-03  9.269791e-04  0.0008978955  9.613805e-04  1.047452e-03
#>  [4,]  3.043311e-04  2.599739e-04  0.0001314316 -1.662837e-05 -4.749907e-05
#>  [5,]  1.172132e-03  9.977547e-04  0.0009652111  1.026426e-03  1.077990e-03
#>  [6,]  1.126081e-03  9.967453e-04  0.0007847704  5.536391e-04  4.612485e-04
#>  [7,] -4.610209e-04 -4.597444e-04 -0.0007601138 -1.095355e-03 -1.164095e-03
#>  [8,]  1.211814e-03  1.005223e-03  0.0009515968  1.007949e-03  1.087662e-03
#>  [9,]  1.502712e-03  1.310035e-03  0.0011254798  9.712600e-04  8.931153e-04
#> [10,] -7.437705e-04 -7.596527e-04 -0.0011421872 -1.635364e-03 -1.754823e-03
#>                 V37           V38          V39            V4          V40
#>  [1,]  2.661659e-04  9.227276e-04 0.0011904048  0.0003132826 8.963293e-04
#>  [2,] -4.107670e-04  3.238457e-04 0.0007155271  0.0002906121 3.916412e-04
#>  [3,]  9.922494e-04  5.973322e-04 0.0003513756 -0.0001589830 5.741397e-04
#>  [4,]  6.605536e-05  3.120248e-04 0.0004325139  0.0001280185 3.288550e-04
#>  [5,]  1.024311e-03  6.329297e-04 0.0003764826 -0.0001273937 6.062606e-04
#>  [6,]  6.096499e-04  8.870270e-04 0.0009756109  0.0001965056 8.516376e-04
#>  [7,] -8.281722e-04 -2.626394e-05 0.0004011174  0.0002827384 8.079168e-05
#>  [8,]  1.056904e-03  6.755619e-04 0.0004280993 -0.0001266430 6.614920e-04
#>  [9,]  9.996087e-04  1.047178e-03 0.0009735796  0.0001136062 9.691288e-04
#> [10,] -1.258134e-03 -7.734075e-05 0.0005791333  0.0003989012 9.028593e-05
#>                V41          V42          V43          V44          V45
#>  [1,] 0.0009991432 0.0010940882 1.101114e-03 0.0011102725 1.288160e-03
#>  [2,] 0.0005750104 0.0008275222 8.409554e-04 0.0008025965 9.261637e-04
#>  [3,] 0.0004135384 0.0001242478 7.536415e-05 0.0001433915 1.504225e-05
#>  [4,] 0.0003721838 0.0004378546 4.379538e-04 0.0004125148 4.244639e-04
#>  [5,] 0.0004575102 0.0001612014 1.179538e-04 0.0001819181 6.767809e-05
#>  [6,] 0.0008623409 0.0008535673 8.289731e-04 0.0008205212 8.728880e-04
#>  [7,] 0.0003070266 0.0006108793 6.572601e-04 0.0006385455 7.759746e-04
#>  [8,] 0.0004988201 0.0001801540 1.260173e-04 0.0002075999 8.313042e-05
#>  [9,] 0.0008954878 0.0007361927 7.060962e-04 0.0007480519 7.628560e-04
#> [10,] 0.0004428286 0.0008824838 9.410992e-04 0.0009216071 1.161603e-03
#>                V46          V47          V48          V49            V5
#>  [1,] 1.017348e-03 7.017431e-04 5.543867e-04 3.280696e-04  3.736088e-04
#>  [2,] 7.316205e-04 5.048072e-04 4.100006e-04 2.416685e-04  3.027372e-04
#>  [3,] 2.609657e-05 9.594069e-05 2.316140e-05 2.539268e-05 -8.335708e-05
#>  [4,] 3.386291e-04 2.553601e-04 1.967814e-04 1.166062e-04  1.567494e-04
#>  [5,] 7.285415e-05 1.269513e-04 4.731826e-05 3.750820e-05 -4.822685e-05
#>  [6,] 6.961423e-04 5.103977e-04 3.900844e-04 2.305804e-04  2.677723e-04
#>  [7,] 6.066129e-04 4.173781e-04 3.508859e-04 2.048810e-04  2.742774e-04
#>  [8,] 7.529795e-05 1.561010e-04 6.665089e-05 4.904466e-05 -4.269458e-05
#>  [9,] 6.130385e-04 4.782202e-04 3.467119e-04 2.054514e-04  2.050567e-04
#> [10,] 9.167856e-04 6.086144e-04 5.048300e-04 2.966315e-04  3.665365e-04
#>                V50           V51           V52           V53          V54
#>  [1,] 1.024114e-04  8.773181e-05  7.151401e-05  3.392186e-05 4.442791e-05
#>  [2,] 6.943953e-05  6.993880e-05  6.014452e-05  2.170960e-05 3.080466e-05
#>  [3,] 1.149239e-05 -2.397000e-06 -9.390665e-06 -1.804235e-06 2.576922e-06
#>  [4,] 3.677715e-05  3.450259e-05  2.874602e-05  1.400984e-05 1.767176e-05
#>  [5,] 1.766190e-05  1.833501e-06 -5.969365e-06  1.606621e-06 5.495521e-06
#>  [6,] 7.667404e-05  6.219851e-05  4.983726e-05  2.797360e-05 3.543344e-05
#>  [7,] 5.324574e-05  6.118486e-05  5.287110e-05  1.435861e-05 2.298849e-05
#>  [8,] 1.935875e-05  3.777152e-06 -4.768072e-06  5.285006e-07 6.507675e-06
#>  [9,] 7.143192e-05  5.144076e-05  3.844153e-05  2.479833e-05 3.206685e-05
#> [10,] 7.726798e-05  8.439362e-05  7.516005e-05  1.692482e-05 2.959621e-05
#>                 V55          V56           V57          V58          V59
#>  [1,]  2.707905e-05 2.790722e-05  1.653471e-05 2.831254e-05 3.031596e-05
#>  [2,]  9.182009e-06 1.256669e-05  8.587465e-07 1.680308e-05 1.773775e-05
#>  [3,]  1.297875e-05 1.089592e-05  1.488211e-05 5.641534e-06 6.447544e-06
#>  [4,]  8.896752e-06 9.262857e-06  5.153222e-06 1.057215e-05 1.126851e-05
#>  [5,]  1.578269e-05 1.316100e-05  1.764578e-05 7.566089e-06 8.363085e-06
#>  [6,]  2.525031e-05 2.434913e-05  1.871913e-05 2.334555e-05 2.472983e-05
#>  [7,]  2.094076e-06 4.569522e-06 -6.943449e-06 1.091282e-05 1.253820e-05
#>  [8,]  1.691407e-05 1.337424e-05  1.810685e-05 8.335564e-06 9.624120e-06
#>  [9,]  2.967613e-05 2.712083e-05  2.494640e-05 2.335629e-05 2.491688e-05
#> [10,] -4.081894e-07 6.148610e-06 -1.148877e-05 1.323937e-05 1.452092e-05
#>                  V6          V60           V7           V8           V9
#>  [1,]  2.954685e-04 3.000862e-05 4.041049e-04 0.0005230846 0.0010039179
#>  [2,]  6.881996e-05 1.958839e-05 1.840266e-04 0.0003489163 0.0007143155
#>  [3,]  2.587936e-04 6.450942e-07 2.437036e-04 0.0001753041 0.0001046092
#>  [4,]  1.106145e-04 1.108427e-05 1.580343e-04 0.0002085949 0.0003592351
#>  [5,]  2.834580e-04 3.130916e-06 2.702059e-04 0.0001965381 0.0001566481
#>  [6,]  3.084368e-04 2.298834e-05 3.783584e-04 0.0004404456 0.0007408250
#>  [7,] -1.760275e-05 1.519013e-05 8.972193e-05 0.0002411841 0.0005709170
#>  [8,]  3.111805e-04 3.371176e-06 3.038552e-04 0.0002197009 0.0001765597
#>  [9,]  3.987615e-04 2.144603e-05 4.452604e-04 0.0004506898 0.0007001073
#> [10,] -8.206336e-05 1.996291e-05 8.147471e-05 0.0003320964 0.0008409842
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]      [,6]
#> [1,]  0.01712011  0.02102311  0.0186082  0.01591226  0.02000736  0.018902
#> [2,] -0.01712011 -0.02102311 -0.0186082 -0.01591226 -0.02000736 -0.018902
#>           [,7]        [,8]        [,9]       [,10]
#> [1,]  0.012409  0.02316613  0.01574769  0.02085669
#> [2,] -0.012409 -0.02316613 -0.01574769 -0.02085669
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5536156 0.4006153 0.7377443 0.1704340 0.8045821 0.3893714 0.1729599
#>  [8] 0.7205807 0.2097888 0.2131825
#> 
#> $B[[2]]
#> [1] -0.04772259 -0.05275207
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  2.882931e-03  1.095858e-03  1.854872e-03  1.119320e-03  2.054531e-03
#>  [6]  2.815439e-03  6.302337e-05  2.118583e-03  3.254972e-03 -1.033301e-04
#> 
#> $vB[[2]]
#> [1]  0.01745491 -0.01745491
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#>  [2,] 0.0209 0.1853 0.1748 0.1556 0.1476 0.1378 0.2584 0.3827 0.4784 0.5360
#>  [3,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#>  [4,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#>  [5,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [6,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#>  [7,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [8,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#>  [9,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [10,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [11,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [12,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [13,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [14,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [15,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [16,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [17,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [18,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [19,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [20,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [21,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [22,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [23,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [24,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [25,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [26,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [27,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [28,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [29,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [30,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [31,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [32,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [33,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [34,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [35,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [36,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [37,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [38,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [39,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [40,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#>  [2,] 0.6192 0.0278 0.7912 0.9264 1.0000 0.9080 0.7435 0.5557 0.3172 0.1295
#>  [3,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#>  [4,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#>  [5,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [6,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#>  [7,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [8,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#>  [9,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [10,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [11,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [12,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [13,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [14,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [15,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [16,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [17,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [18,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [19,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [20,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [21,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [22,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [23,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [24,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [25,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [26,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [27,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [28,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [29,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [30,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [31,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [32,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [33,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [34,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [35,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [36,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [37,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [38,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [39,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [40,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#>  [2,] 0.0598 0.2722 0.0115 0.3616 0.3293 0.4855 0.3936 0.1845 0.0342 0.2489
#>  [3,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#>  [4,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#>  [5,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [6,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#>  [7,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [8,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#>  [9,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [10,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [11,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [12,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [13,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [14,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [15,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [16,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [17,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [18,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [19,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [20,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [21,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [22,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [23,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [24,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [25,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [26,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [27,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [28,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [29,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [30,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [31,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [32,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [33,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [34,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [35,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [36,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [37,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [38,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [39,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [40,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#>  [2,] 0.3837 0.3514 0.2654 0.0445 0.1760 0.1599 0.0866 0.0590 0.0813 0.0492
#>  [3,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#>  [4,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#>  [5,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [6,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#>  [7,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [8,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#>  [9,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [10,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [11,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [12,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [13,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [14,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [15,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [16,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [17,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [18,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [19,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [20,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [21,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [22,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [23,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [24,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [25,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [26,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [27,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [28,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [29,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [30,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [31,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [32,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [33,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [34,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [35,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [36,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [37,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [38,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [39,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [40,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#>  [2,] 0.0417 0.0495 0.0367 0.0115 0.0427 0.0118 0.0133 0.0096 0.0014 0.0049
#>  [3,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#>  [4,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#>  [5,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [6,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#>  [7,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [8,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#>  [9,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [10,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [11,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [12,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [13,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [14,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [15,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [16,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [17,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [18,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [19,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [20,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [21,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [22,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [23,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [24,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [25,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [26,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [27,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [28,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [29,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [30,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [31,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [32,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [33,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [34,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [35,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [36,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [37,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [38,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [39,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [40,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#>  [2,] 0.0039 0.0029 0.0078 0.0047 0.0021 0.0766 0.0011 0.1458 0.1430 0.1894
#>  [3,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#>  [4,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#>  [5,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [6,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#>  [7,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [8,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#>  [9,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [10,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [11,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [12,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [13,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [14,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [15,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [16,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [17,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [18,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [19,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [20,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [21,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [22,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [23,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [24,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [25,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [26,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [27,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [28,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [29,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [30,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [31,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [32,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [33,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [34,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [35,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [36,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [37,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [38,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [39,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [40,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5939276 0.4493057 0.9122314 0.3150247 0.8897555 0.5850780 0.2797157
#>  [2,] 0.6491607 0.4492635 0.8736506 0.3868854 0.8421427 0.5634790 0.3157628
#>  [3,] 0.6214388 0.4048158 0.8916726 0.3337243 0.8659917 0.5879148 0.2840073
#>  [4,] 0.5881122 0.2677029 0.7441769 0.2804943 0.7468628 0.3924624 0.2782033
#>  [5,] 0.6118647 0.4613798 0.9215351 0.3138535 0.8963142 0.6124776 0.2713301
#>  [6,] 0.5620236 0.4275542 0.8555944 0.2884550 0.8451850 0.5347843 0.3053517
#>  [7,] 0.7628393 0.4362520 0.9301544 0.4330144 0.9064275 0.6490914 0.2770759
#>  [8,] 0.7176704 0.3930069 0.9292186 0.3793753 0.8965372 0.6334209 0.2528841
#>  [9,] 0.7532058 0.4351777 0.9289096 0.4323659 0.9041272 0.6432768 0.2784279
#> [10,] 0.6941302 0.4153816 0.9231881 0.3740105 0.8815648 0.6227682 0.2614515
#> [11,] 0.5943099 0.3038961 0.8751438 0.2748224 0.8474377 0.5331913 0.1993840
#> [12,] 0.5798376 0.2847496 0.8870993 0.2463124 0.8724703 0.4661989 0.1778219
#> [13,] 0.4533097 0.2804154 0.8783432 0.1658559 0.8580938 0.4138311 0.1703700
#> [14,] 0.6586945 0.4916048 0.9447527 0.3835028 0.9159645 0.6679548 0.2774456
#> [15,] 0.5564699 0.3794041 0.9121217 0.2967543 0.8847449 0.5716722 0.2214300
#> [16,] 0.6136483 0.2622518 0.8199322 0.2545500 0.8342016 0.4302614 0.2476728
#> [17,] 0.3495346 0.2099847 0.8150542 0.1158921 0.8108109 0.3173141 0.1455290
#> [18,] 0.6779907 0.3799456 0.9130935 0.3463178 0.8886763 0.6171570 0.2379211
#> [19,] 0.5678475 0.2849637 0.8624680 0.2644682 0.8319292 0.4975716 0.1838869
#> [20,] 0.6044628 0.3219085 0.8827655 0.2785257 0.8547592 0.5381005 0.2021184
#> [21,] 0.6690482 0.3940228 0.8816038 0.3584139 0.8535840 0.5635657 0.2855882
#> [22,] 0.7455229 0.4468553 0.9159145 0.4382300 0.8905494 0.6432379 0.2947878
#> [23,] 0.5160657 0.2107764 0.8636851 0.1490577 0.8368835 0.3701091 0.1619199
#> [24,] 0.5888617 0.4001220 0.9299853 0.2979582 0.9035066 0.6118400 0.2194002
#> [25,] 0.6490393 0.4843465 0.8656932 0.4374007 0.8455197 0.6294586 0.3482227
#> [26,] 0.7031253 0.3814645 0.9166347 0.3766989 0.8859140 0.6120732 0.2520370
#> [27,] 0.5689913 0.3054161 0.9102931 0.2476105 0.8778677 0.5067859 0.2041833
#> [28,] 0.4978916 0.2960089 0.8669186 0.2070767 0.8183438 0.4455753 0.1858810
#> [29,] 0.6828232 0.3102122 0.8974214 0.2967382 0.8798782 0.5227438 0.2338884
#> [30,] 0.4579109 0.2366421 0.7438417 0.1962033 0.7426590 0.4117317 0.1956396
#> [31,] 0.6911742 0.3518320 0.9044536 0.3283231 0.8696523 0.5716254 0.2468786
#> [32,] 0.3255523 0.2232782 0.8306669 0.1144989 0.8094279 0.3172272 0.1352049
#> [33,] 0.3931840 0.3615010 0.8700713 0.2180979 0.8651633 0.4718867 0.2139767
#> [34,] 0.5553875 0.3718691 0.8819893 0.2931502 0.8592254 0.5400382 0.2393532
#> [35,] 0.6808898 0.3972812 0.8975364 0.3686777 0.8657510 0.5853630 0.2839298
#> [36,] 0.5952724 0.4046492 0.9407473 0.3095145 0.9115760 0.6188600 0.2089176
#> [37,] 0.6565393 0.4541549 0.9111257 0.3839917 0.8864767 0.6351258 0.2966293
#> [38,] 0.5604548 0.3448077 0.9054980 0.2490079 0.8859439 0.5296044 0.2361347
#> [39,] 0.6155459 0.2377637 0.8459821 0.2436839 0.8281001 0.4385158 0.2042483
#> [40,] 0.4965957 0.2621888 0.8772107 0.1772785 0.8604887 0.4603473 0.1672388
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8856755 0.6628280 0.4086411
#>  [2,] 0.8287910 0.6369346 0.3405436
#>  [3,] 0.8693588 0.6574123 0.3567379
#>  [4,] 0.6375567 0.4182555 0.2218471
#>  [5,] 0.8963217 0.6914734 0.4006862
#>  [6,] 0.8237533 0.5835939 0.4075792
#>  [7,] 0.9053342 0.7028724 0.3074054
#>  [8,] 0.9005907 0.6889889 0.2864230
#>  [9,] 0.9035198 0.6996457 0.3096615
#> [10,] 0.8919750 0.6971242 0.3009905
#> [11,] 0.8186578 0.5665020 0.2391626
#> [12,] 0.8338398 0.5347301 0.2214362
#> [13,] 0.8050086 0.5546510 0.2451581
#> [14,] 0.9234966 0.7414481 0.4095901
#> [15,] 0.8761587 0.6512823 0.3300384
#> [16,] 0.7613722 0.4906411 0.2403191
#> [17,] 0.6681228 0.4300124 0.1949884
#> [18,] 0.8816552 0.6534792 0.2920518
#> [19,] 0.8038061 0.5302622 0.2254642
#> [20,] 0.8432719 0.5811855 0.2576867
#> [21,] 0.8318968 0.6242311 0.2999368
#> [22,] 0.8913949 0.6943643 0.3172036
#> [23,] 0.7703786 0.4767826 0.1656096
#> [24,] 0.8957407 0.6504983 0.3414632
#> [25,] 0.8525622 0.6426739 0.4222567
#> [26,] 0.8903455 0.6667750 0.2799706
#> [27,] 0.8572647 0.6340748 0.2407232
#> [28,] 0.8136790 0.5796919 0.2443247
#> [29,] 0.8642002 0.6137703 0.2508558
#> [30,] 0.6851452 0.4210851 0.2232640
#> [31,] 0.8629419 0.6422842 0.2590894
#> [32,] 0.6735480 0.4462306 0.1948540
#> [33,] 0.8185078 0.5420757 0.3742542
#> [34,] 0.8283380 0.6006587 0.3216281
#> [35,] 0.8821714 0.6564854 0.3215174
#> [36,] 0.9056559 0.6680143 0.3320598
#> [37,] 0.8925301 0.6884651 0.3824834
#> [38,] 0.8438679 0.6124278 0.3134689
#> [39,] 0.7829225 0.5076652 0.1966031
#> [40,] 0.8279029 0.5713291 0.2534629
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5570237 0.4429763
#>  [2,] 0.5527573 0.4472427
#>  [3,] 0.5575409 0.4424591
#>  [4,] 0.5502281 0.4497719
#>  [5,] 0.5581147 0.4418853
#>  [6,] 0.5515935 0.4484065
#>  [7,] 0.5606191 0.4393809
#>  [8,] 0.5636410 0.4363590
#>  [9,] 0.5604423 0.4395577
#> [10,] 0.5616756 0.4383244
#> [11,] 0.5650561 0.4349439
#> [12,] 0.5690310 0.4309690
#> [13,] 0.5691037 0.4308963
#> [14,] 0.5582041 0.4417959
#> [15,] 0.5637874 0.4362126
#> [16,] 0.5601502 0.4398498
#> [17,] 0.5667692 0.4332308
#> [18,] 0.5632893 0.4367107
#> [19,] 0.5653839 0.4346161
#> [20,] 0.5649625 0.4350375
#> [21,] 0.5573771 0.4426229
#> [22,] 0.5580602 0.4419398
#> [23,] 0.5711277 0.4288723
#> [24,] 0.5642447 0.4357553
#> [25,] 0.5473986 0.4526014
#> [26,] 0.5631902 0.4368098
#> [27,] 0.5692800 0.4307200
#> [28,] 0.5660362 0.4339638
#> [29,] 0.5657954 0.4342046
#> [30,] 0.5567311 0.4432689
#> [31,] 0.5636382 0.4363618
#> [32,] 0.5678601 0.4321399
#> [33,] 0.5607667 0.4392333
#> [34,] 0.5602246 0.4397754
#> [35,] 0.5588296 0.4411704
#> [36,] 0.5657325 0.4342675
#> [37,] 0.5562490 0.4437510
#> [38,] 0.5639398 0.4360602
#> [39,] 0.5654239 0.4345761
#> [40,] 0.5693539 0.4306461
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]        [,2]     [,3]       [,4]     [,5]         [,6]
#>  [1,]  0.380226128 -0.20347632 2.341190 -0.7767316 2.088246  0.343654601
#>  [2,]  0.615351915 -0.20364678 1.933629 -0.4604234 1.674258  0.255293702
#>  [3,]  0.495659840 -0.38543881 2.107940 -0.6913884 1.865974  0.355351762
#>  [4,]  0.356166782 -1.00630831 1.067792 -0.9420108 1.081950 -0.436973424
#>  [5,]  0.455157304 -0.15478908 2.463390 -0.7821651 2.156926  0.457738622
#>  [6,]  0.249378897 -0.29183682 1.779171 -0.9028996 1.697324  0.139362423
#>  [7,]  1.168309184 -0.25638728 2.589063 -0.2695629 2.270775  0.615047531
#>  [8,]  0.932935115 -0.43469039 2.574748 -0.4922004 2.159328  0.546920598
#>  [9,]  1.115783954 -0.26075664 2.570059 -0.2722047 2.243948  0.589615200
#> [10,]  0.819500215 -0.34176160 2.486474 -0.5150498 2.007333  0.501314360
#> [11,]  0.381811344 -0.82881323 1.947225 -0.9702917 1.714645  0.132960739
#> [12,]  0.322106788 -0.92102238 2.061449 -1.1183773 1.922980 -0.135410765
#> [13,] -0.187307048 -0.94240201 1.976833 -1.6152865 1.799547 -0.348150195
#> [14,]  0.657481731 -0.03358397 2.839104 -0.4747067 2.388738  0.698949209
#> [15,]  0.226847541 -0.49207826 2.339821 -0.8628018 2.038151  0.288676860
#> [16,]  0.462673826 -1.03429727 1.515888 -1.0744910 1.615703 -0.280784880
#> [17,] -0.621085604 -1.32501734 1.483192 -2.0319199 1.455288 -0.766142962
#> [18,]  0.744553000 -0.48977915 2.352005 -0.6352646 2.077291  0.477498262
#> [19,]  0.273074438 -0.91997173 1.835942 -1.0228725 1.599362 -0.009713539
#> [20,]  0.424095183 -0.74501476 2.018883 -0.9517866 1.772427  0.152698023
#> [21,]  0.703883424 -0.43043340 2.007706 -0.5822551 1.762992  0.255646208
#> [22,]  1.074875224 -0.21338499 2.388089 -0.2483486 2.096366  0.589445566
#> [23,]  0.064285070 -1.32025155 1.846241 -1.7420110 1.635220 -0.531748894
#> [24,]  0.359261632 -0.40495683 2.586463 -0.8570399 2.236809  0.455053016
#> [25,]  0.614818998 -0.06263457 1.863404 -0.2517181 1.699885  0.529894837
#> [26,]  0.862224948 -0.48333681 2.397477 -0.5035834 2.049667  0.456035173
#> [27,]  0.277737069 -0.82163783 2.317219 -1.1113972 1.972391  0.027145319
#> [28,] -0.008433712 -0.86637613 1.873984 -1.3426372 1.505167 -0.218564773
#> [29,]  0.766776392 -0.79912745 2.168896 -0.8628790 1.991277  0.091038061
#> [30,] -0.168755823 -1.17117792 1.066033 -1.4101951 1.059835 -0.356811370
#> [31,]  0.805614623 -0.61099620 2.247719 -0.7157793 1.897888  0.288485790
#> [32,] -0.728370827 -1.24666370 1.590361 -2.0455881 1.446297 -0.766544086
#> [33,] -0.433947978 -0.56885547 1.901589 -1.2767857 1.858854 -0.112571915
#> [34,]  0.222462950 -0.52420679 2.011404 -0.8801330 1.808872  0.160496513
#> [35,]  0.757864163 -0.41680628 2.170146 -0.5378938 1.863901  0.344828736
#> [36,]  0.385804971 -0.38613043 2.764862 -0.8023898 2.333031  0.484712414
#> [37,]  0.647909911 -0.18389673 2.327458 -0.4726395 2.055247  0.554270637
#> [38,]  0.243008006 -0.64194207 2.259864 -1.1039106 2.049963  0.118556257
#> [39,]  0.470684987 -1.16497912 1.703429 -1.1325875 1.572222 -0.247187723
#> [40,] -0.013617515 -1.03462333 1.966277 -1.5348956 1.819354 -0.158944532
#>             [,7]      [,8]        [,9]      [,10]
#>  [1,] -0.9458722 2.0473102  0.67592223 -0.3695857
#>  [2,] -0.7733131 1.5770832  0.56208388 -0.6608728
#>  [3,] -0.9246698 1.8953009  0.65178370 -0.5895508
#>  [4,] -0.9533913 0.5647750 -0.32993899 -1.2549346
#>  [5,] -0.9878847 2.1570065  0.80701641 -0.4026067
#>  [6,] -0.8219417 1.5419864  0.33754438 -0.3739820
#>  [7,] -0.9590127 2.2579509  0.86101375 -0.8122772
#>  [8,] -1.0832888 2.2038051  0.79539661 -0.9128205
#>  [9,] -0.9522732 2.2369601  0.84561119 -0.8017024
#> [10,] -1.0384382 2.1110753  0.83364072 -0.8425858
#> [11,] -1.3901488 1.5072802  0.26759357 -1.1572759
#> [12,] -1.5311743 1.6130886  0.13914428 -1.2573165
#> [13,] -1.5830071 1.4178976  0.21948096 -1.1246052
#> [14,] -0.9571678 2.4908325  1.05350863 -0.3656601
#> [15,] -1.2573526 1.9565461  0.62468028 -0.7080112
#> [16,] -1.1110628 1.1602174 -0.03743983 -1.1509307
#> [17,] -1.7701070 0.6997070 -0.28180042 -1.4179169
#> [18,] -1.1641110 2.0081982  0.63436768 -0.8854399
#> [19,] -1.4902322 1.4102546  0.12119698 -1.2341024
#> [20,] -1.3731063 1.6827769  0.32764168 -1.0580271
#> [21,] -0.9169087 1.5991304  0.50754612 -0.8475987
#> [22,] -0.8722429 2.1050694  0.82060280 -0.7666531
#> [23,] -1.6440116 1.2104501 -0.09293658 -1.6170682
#> [24,] -1.2691650 2.1507703  0.62123047 -0.6567805
#> [25,] -0.6268609 1.7548397  0.58698873 -0.3135161
#> [26,] -1.0877778 2.0942750  0.69363485 -0.9446076
#> [27,] -1.3603510 1.7927552  0.54973797 -1.1487186
#> [28,] -1.4770001 1.4740948  0.32150886 -1.1291136
#> [29,] -1.1864835 1.8506229  0.46318838 -1.0940530
#> [30,] -1.4137732 0.7775192 -0.31832060 -1.2467455
#> [31,] -1.1153299 1.8399425  0.58529197 -1.0507067
#> [32,] -1.8557012 0.7242762 -0.21591268 -1.4187732
#> [33,] -1.3011193 1.5062700  0.16870178 -0.5140090
#> [34,] -1.1562286 1.5738936  0.40821032 -0.7462998
#> [35,] -0.9250510 2.0131556  0.64767096 -0.7468071
#> [36,] -1.3314622 2.2617105  0.69921777 -0.6988834
#> [37,] -0.8634010 2.1168497  0.79295328 -0.4790208
#> [38,] -1.1739888 1.6872939  0.45752876 -0.7839513
#> [39,] -1.3599510 1.2827794  0.03066302 -1.4076616
#> [40,] -1.6053243 1.5708373  0.28727599 -1.0802278
#> 
#> $pre[[3]]
#>                [,1]       [,2]
#>  [1,] -0.0335247626 -0.2626163
#>  [2,] -0.0456550426 -0.2574725
#>  [3,] -0.0315558164 -0.2627435
#>  [4,] -0.0197123602 -0.2213047
#>  [5,] -0.0329756950 -0.2664898
#>  [6,] -0.0385239419 -0.2456352
#>  [7,] -0.0361244227 -0.2797995
#>  [8,] -0.0226566309 -0.2786087
#>  [9,] -0.0362137168 -0.2791710
#> [10,] -0.0271954936 -0.2751607
#> [11,]  0.0004484625 -0.2612594
#> [12,]  0.0174989533 -0.2603997
#> [13,]  0.0263617106 -0.2518336
#> [14,] -0.0435712338 -0.2774477
#> [15,] -0.0122346201 -0.2687820
#> [16,]  0.0009029619 -0.2408685
#> [17,]  0.0361119611 -0.2325694
#> [18,] -0.0188305427 -0.2733529
#> [19,]  0.0055224803 -0.2575193
#> [20,] -0.0013413878 -0.2626684
#> [21,] -0.0301977868 -0.2607217
#> [22,] -0.0425610308 -0.2758541
#> [23,]  0.0396246520 -0.2468289
#> [24,] -0.0147325896 -0.2731399
#> [25,] -0.0688682425 -0.2590338
#> [26,] -0.0213522139 -0.2754716
#> [27,]  0.0123830509 -0.2665312
#> [28,]  0.0131859333 -0.2525111
#> [29,]  0.0004949167 -0.2642217
#> [30,]  0.0012056341 -0.2267001
#> [31,] -0.0122555838 -0.2681966
#> [32,]  0.0375045352 -0.2356213
#> [33,] -0.0058632929 -0.2501376
#> [34,] -0.0168136192 -0.2588873
#> [35,] -0.0302179375 -0.2666314
#> [36,] -0.0128359273 -0.2772965
#> [37,] -0.0438167365 -0.2697693
#> [38,] -0.0037445801 -0.2609117
#> [39,]  0.0138729919 -0.2493318
#> [40,]  0.0236493315 -0.2555661
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5570237  0.5570237
#>  [2,]  0.4472427 -0.4472427
#>  [3,]  0.4424591 -0.4424591
#>  [4,]  0.4497719 -0.4497719
#>  [5,] -0.5581147  0.5581147
#>  [6,] -0.5515935  0.5515935
#>  [7,] -0.5606191  0.5606191
#>  [8,]  0.4363590 -0.4363590
#>  [9,] -0.5604423  0.5604423
#> [10,]  0.4383244 -0.4383244
#> [11,]  0.4349439 -0.4349439
#> [12,] -0.5690310  0.5690310
#> [13,] -0.5691037  0.5691037
#> [14,] -0.5582041  0.5582041
#> [15,] -0.5637874  0.5637874
#> [16,] -0.5601502  0.5601502
#> [17,]  0.4332308 -0.4332308
#> [18,]  0.4367107 -0.4367107
#> [19,] -0.5653839  0.5653839
#> [20,] -0.5649625  0.5649625
#> [21,] -0.5573771  0.5573771
#> [22,] -0.5580602  0.5580602
#> [23,] -0.5711277  0.5711277
#> [24,]  0.4357553 -0.4357553
#> [25,] -0.5473986  0.5473986
#> [26,]  0.4368098 -0.4368098
#> [27,]  0.4307200 -0.4307200
#> [28,]  0.4339638 -0.4339638
#> [29,] -0.5657954  0.5657954
#> [30,]  0.4432689 -0.4432689
#> [31,]  0.4363618 -0.4363618
#> [32,]  0.4321399 -0.4321399
#> [33,] -0.5607667  0.5607667
#> [34,] -0.5602246  0.5602246
#> [35,] -0.5588296  0.5588296
#> [36,]  0.4342675 -0.4342675
#> [37,] -0.5562490  0.5562490
#> [38,]  0.4360602 -0.4360602
#> [39,]  0.4345761 -0.4345761
#> [40,]  0.4306461 -0.4306461
#> 
#> $L
#> [1] 0.6866225 0.7820644 0.9112340 0.6906843 0.7471955 0.7047894
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3478261