Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11         V12          V13
#>  [1,] -0.3148357 -0.166849057 -0.163581879 -0.05781477 -0.051433045
#>  [2,] -0.2520076  0.008699202 -0.187925000 -0.13975421 -0.051161638
#>  [3,] -0.2605936 -0.187561840 -0.155468295 -0.17499703 -0.047135798
#>  [4,] -0.3607539 -0.040761959 -0.004643522  0.03801697 -0.004123485
#>  [5,] -0.3540899 -0.165041554 -0.099412421 -0.17369353 -0.095553144
#>  [6,] -0.3175146 -0.137672948 -0.028779610  0.01057411 -0.080063414
#>  [7,] -0.3038748 -0.128235971 -0.039426427 -0.16979712  0.017255549
#>  [8,] -0.3438560 -0.059117743 -0.048737472 -0.04690259 -0.162546919
#>  [9,] -0.3634128 -0.133634160 -0.220228216 -0.12263273 -0.140990737
#> [10,] -0.3191590 -0.076916994  0.011155917 -0.10578677 -0.048841876
#>                V14           V15         V16         V17         V18
#>  [1,] -0.141878014 -0.1904685759  0.02946993  0.04271074  0.06069013
#>  [2,] -0.171198606 -0.0450107711 -0.11796506 -0.13606743 -0.04930560
#>  [3,] -0.076856505 -0.0503430586 -0.09490476 -0.09315427 -0.18546207
#>  [4,]  0.015688380 -0.1146601158  0.06093408 -0.02700831 -0.07862654
#>  [5,] -0.211537945 -0.1760390487 -0.13037981 -0.13306387 -0.08377279
#>  [6,] -0.000126328  0.0888325501  0.11175281  0.18573788  0.11685916
#>  [7,]  0.061158533 -0.1157753456  0.06234505 -0.01939530  0.01049760
#>  [8,] -0.170073018 -0.0004926965 -0.13113792 -0.05947780  0.03244761
#>  [9,] -0.135934383 -0.1498779181 -0.15094097 -0.08022953 -0.01245048
#> [10,]  0.063874973  0.0909082761  0.06499617  0.11793595  0.18288305
#>               V19         V2         V20          V21         V22         V23
#>  [1,] -0.03826338 -0.2838453  0.13639999  0.193576997  0.22649992  0.07202829
#>  [2,] -0.10941474 -0.2507095 -0.05839887 -0.017135691  0.10210315  0.05417201
#>  [3,] -0.02624873 -0.2349428  0.03202762  0.131163834 -0.07834200 -0.02747647
#>  [4,] -0.04127697 -0.2225433  0.09166381  0.088311619  0.04766474  0.11925866
#>  [5,] -0.05633995 -0.2488898 -0.08200546  0.007900594 -0.07639782  0.07013759
#>  [6,]  0.17102096 -0.4028251  0.19609874  0.180116133  0.28605412  0.28593169
#>  [7,]  0.12888429 -0.2973497 -0.04069758  0.088020641  0.20114027  0.22234531
#>  [8,]  0.04329859 -0.2405730 -0.03784195  0.007939870 -0.01283353 -0.07489918
#>  [9,]  0.10854175 -0.3062640  0.10835702  0.094912563  0.10158479  0.08770619
#> [10,]  0.04726140 -0.3735419  0.18088316  0.072533738  0.12467678  0.12440373
#>               V24           V25           V26        V27         V28        V29
#>  [1,]  0.16539303  1.679690e-01  0.2145445543 0.07437056  0.08487212 0.12604851
#>  [2,]  0.01173901 -1.971698e-05  0.0894913888 0.12841143  0.11937870 0.01247481
#>  [3,]  0.07755814  1.272791e-01 -0.0001854162 0.18224011  0.07891505 0.19173820
#>  [4,]  0.08149233  1.541710e-01  0.1813247325 0.14490179  0.14768932 0.12741589
#>  [5,]  0.00527470  1.456968e-01  0.1584409445 0.14590757 -0.02158783 0.03658072
#>  [6,]  0.25310599  2.790984e-01  0.1540354745 0.22228975  0.26435482 0.01798927
#>  [7,]  0.19647032  1.613698e-01  0.2125869852 0.22960955  0.14915849 0.07764687
#>  [8,] -0.07631866  2.364204e-02  0.1229578082 0.08698490  0.07909697 0.02846325
#>  [9,]  0.19263104  1.037579e-01  0.1041206706 0.08271574  0.12835651 0.12366191
#> [10,]  0.24963132  1.829686e-01  0.2827513619 0.11580152  0.13701553 0.06808605
#>               V3         V30          V31          V32          V33
#>  [1,] -0.2730528  0.02783117 -0.082169134 -0.106648006 -0.007650542
#>  [2,] -0.2698621  0.12338808  0.005695021  0.098560092 -0.068328819
#>  [3,] -0.2392714  0.10194268  0.116057914  0.073737971 -0.055366855
#>  [4,] -0.3683156  0.02090062  0.046673157  0.085047323  0.100185546
#>  [5,] -0.3301798 -0.05660173 -0.040363078 -0.095018970 -0.166520510
#>  [6,] -0.4321241  0.02817964  0.059525965  0.010485732  0.034871461
#>  [7,] -0.2366558 -0.02696728  0.008283977 -0.098902412 -0.056466070
#>  [8,] -0.2358872  0.09921430 -0.096144989  0.029877242 -0.096827182
#>  [9,] -0.2510515  0.15533859  0.055426591 -0.004782832 -0.121072975
#> [10,] -0.2909200  0.13041535  0.160986499  0.045794957  0.084918587
#>                V34          V35          V36          V37         V38
#>  [1,] -0.037510593 -0.152393827 -0.039060849 -0.015249267 -0.02220194
#>  [2,] -0.040435469 -0.105388924 -0.018471877 -0.103055483  0.02909047
#>  [3,]  0.044074130 -0.043403678 -0.042732181  0.058965421 -0.09220036
#>  [4,] -0.076051915 -0.025395341 -0.057186413 -0.021000914  0.08403635
#>  [5,] -0.050122918  0.039722414 -0.128683060 -0.006438207  0.00778694
#>  [6,] -0.007158701 -0.034845099  0.002780256 -0.059853284  0.05580557
#>  [7,] -0.042342180 -0.041581567 -0.038388684 -0.035585343 -0.15278235
#>  [8,]  0.029495985 -0.003105182 -0.111861227 -0.098970104 -0.12699480
#>  [9,]  0.005956120 -0.037993411 -0.023184521  0.078797348  0.01736948
#> [10,] -0.113790864 -0.022686350  0.083629007  0.033585049 -0.01441578
#>                V39         V4          V40          V41          V42
#>  [1,] -0.096229854 -0.3619145 -0.036290595 -0.193742273 -0.016545365
#>  [2,]  0.065085311 -0.1611062 -0.137920586 -0.078391100 -0.177346203
#>  [3,] -0.059100149 -0.1906378 -0.101752700 -0.081453753  0.027119800
#>  [4,]  0.004698008 -0.3263983  0.008426243  0.041368992 -0.038201795
#>  [5,] -0.134094450 -0.2223261 -0.041475690 -0.205709897 -0.190422978
#>  [6,] -0.027526014 -0.2708361 -0.077328359 -0.031680255 -0.142812448
#>  [7,] -0.083833874 -0.3372515  0.036275556 -0.097627520 -0.066895768
#>  [8,] -0.051188609 -0.2849044 -0.018720083 -0.165927126 -0.134759766
#>  [9,]  0.082066552 -0.3206082  0.042116616 -0.025356771 -0.003160681
#> [10,]  0.105137790 -0.2841460 -0.054513402  0.007646834 -0.057996040
#>               V43         V44         V45         V46         V47        V48
#>  [1,] -0.18368375 -0.08259736 -0.14086971 -0.20391551 -0.25792331 -0.1555292
#>  [2,] -0.14868109 -0.11467911 -0.17481213 -0.15910874 -0.17217874 -0.2615697
#>  [3,]  0.02049978 -0.07042678 -0.07229899 -0.13969441 -0.16169683 -0.3092931
#>  [4,] -0.01526053 -0.14363123 -0.12194231 -0.08032860 -0.12060990 -0.1522562
#>  [5,] -0.12792539 -0.08518107 -0.15494282 -0.13650525 -0.13431719 -0.1423275
#>  [6,] -0.02505987 -0.20055403 -0.04139366 -0.06080645 -0.24060168 -0.2782783
#>  [7,] -0.05998246 -0.13748906 -0.07470358 -0.27880434 -0.27834808 -0.2594267
#>  [8,] -0.19092431 -0.07725874 -0.09957414 -0.13931281 -0.22282568 -0.1807137
#>  [9,] -0.13056302 -0.07819519 -0.09995735 -0.24204584 -0.11520630 -0.3053137
#> [10,] -0.07094673  0.00104996 -0.03666445 -0.05864936 -0.06447258 -0.1116416
#>              V49         V5        V50        V51        V52        V53
#>  [1,] -0.2255737 -0.2722637 -0.3944661 -0.2472261 -0.3323756 -0.4023119
#>  [2,] -0.1994374 -0.1877874 -0.2208166 -0.2905454 -0.1920357 -0.1799960
#>  [3,] -0.1982331 -0.1858696 -0.3635120 -0.3820918 -0.2705329 -0.2649289
#>  [4,] -0.3299768 -0.2527613 -0.3588597 -0.2846968 -0.4071914 -0.3022429
#>  [5,] -0.3032757 -0.2154940 -0.3393964 -0.2898377 -0.2811558 -0.3015505
#>  [6,] -0.3851612 -0.2710890 -0.3634339 -0.4626082 -0.4370042 -0.4512221
#>  [7,] -0.3736470 -0.2922025 -0.3586031 -0.4171696 -0.2817597 -0.3498693
#>  [8,] -0.3365251 -0.2502412 -0.3648883 -0.3311436 -0.3162035 -0.2611720
#>  [9,] -0.3362530 -0.2874162 -0.2190678 -0.2622927 -0.2674228 -0.3209705
#> [10,] -0.2724315 -0.1766987 -0.3635556 -0.3243398 -0.4032667 -0.3802818
#>              V54        V55        V56        V57        V58        V59
#>  [1,] -0.4219447 -0.3614027 -0.2910692 -0.4411480 -0.2699357 -0.3384631
#>  [2,] -0.1839650 -0.3218500 -0.1751620 -0.2560858 -0.1686445 -0.2034048
#>  [3,] -0.3147136 -0.3498118 -0.4082371 -0.3324684 -0.3932120 -0.3358709
#>  [4,] -0.4174331 -0.2606841 -0.3124363 -0.4375121 -0.3525973 -0.2797371
#>  [5,] -0.2156388 -0.3366638 -0.2432631 -0.2509241 -0.3672205 -0.2682709
#>  [6,] -0.3588863 -0.3617215 -0.4114034 -0.4070377 -0.3881858 -0.5343327
#>  [7,] -0.4641522 -0.4122343 -0.4300130 -0.4574706 -0.4321619 -0.3758323
#>  [8,] -0.2693264 -0.3971133 -0.2595420 -0.2382528 -0.3681857 -0.3114608
#>  [9,] -0.2590481 -0.3806477 -0.2889378 -0.3644923 -0.3048751 -0.3337158
#> [10,] -0.2645858 -0.3147319 -0.3960577 -0.2792418 -0.2609310 -0.2582583
#>               V6        V60          V7         V8          V9
#>  [1,] -0.1617905 -0.3837832 -0.16184763 -0.1767579 -0.15135721
#>  [2,] -0.2219652 -0.1840835 -0.14162866 -0.2143156 -0.06733805
#>  [3,] -0.2366380 -0.2366184 -0.18411935 -0.2668754 -0.19733574
#>  [4,] -0.1552347 -0.3638005 -0.14064245 -0.1358265 -0.13561561
#>  [5,] -0.1431716 -0.3388681 -0.18680700 -0.1033321 -0.14169296
#>  [6,] -0.2217518 -0.3957006 -0.22870196 -0.2349068 -0.13027895
#>  [7,] -0.2422572 -0.4502019 -0.27798738 -0.2163012 -0.20599719
#>  [8,] -0.2709234 -0.3370856 -0.29561670 -0.1623246 -0.17624942
#>  [9,] -0.1665945 -0.2746619 -0.17753131 -0.2224382 -0.13364340
#> [10,] -0.1708112 -0.3708733 -0.08485598 -0.1242156 -0.17458219
#> 
#> $W[[2]]
#>             [,1]       [,2]         [,3]       [,4]       [,5]        [,6]
#> [1,]  0.01954806 0.03640268 -0.002928638 0.03343737 -0.0124798  0.10227884
#> [2,] -0.08540932 0.01327101  0.009299672 0.04034726  0.0723240 -0.08209761
#>             [,7]       [,8]         [,9]       [,10]
#> [1,]  0.01101878 0.04265101 -0.099138966 0.029214355
#> [2,] -0.05924886 0.10683818  0.005941089 0.009630512
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -2.968656e-05 -0.0002182955 -3.594602e-04 -0.0002075987 -5.090824e-06
#>  [2,]  2.117317e-05  0.0001307742  8.274856e-05  0.0001657320  2.956229e-04
#>  [3,]  1.143142e-04  0.0008781855  9.881312e-04  0.0010029950  1.052465e-03
#>  [4,]  1.770923e-04  0.0013623453  1.556874e-03  0.0014837390  1.477166e-03
#>  [5,]  1.505986e-04  0.0012208682  1.490805e-03  0.0014500484  1.348313e-03
#>  [6,] -5.459778e-05 -0.0003756188 -4.569971e-04 -0.0002375273 -1.050367e-04
#>  [7,]  3.824984e-05  0.0003177303  2.949700e-04  0.0003832132  5.033753e-04
#>  [8,]  1.694019e-04  0.0013501832  1.613617e-03  0.0015557252  1.482925e-03
#>  [9,]  2.967195e-04  0.0022832252  2.690679e-03  0.0025293983  2.402869e-03
#> [10,]  1.312751e-04  0.0010151763  1.128100e-03  0.0010807000  1.082834e-03
#>                 V14          V15          V16          V17          V18
#>  [1,]  2.258804e-04 0.0004977551 0.0008678589 0.0010665857 0.0009719093
#>  [2,]  4.139989e-04 0.0005661658 0.0008516406 0.0009976230 0.0009190815
#>  [3,]  1.052794e-03 0.0010370842 0.0012751350 0.0013926809 0.0013398369
#>  [4,]  1.453530e-03 0.0013434827 0.0014812208 0.0015257820 0.0014967653
#>  [5,]  1.134138e-03 0.0009182719 0.0009459108 0.0009888476 0.0010541952
#>  [6,] -1.714981e-05 0.0001327479 0.0003502173 0.0005103467 0.0005032512
#>  [7,]  6.247845e-04 0.0007456737 0.0010310111 0.0011791886 0.0011115337
#>  [8,]  1.320167e-03 0.0011183926 0.0011804097 0.0012211386 0.0012594509
#>  [9,]  2.178903e-03 0.0018136154 0.0018458922 0.0018312763 0.0018657221
#> [10,]  1.091287e-03 0.0010413254 0.0011655348 0.0012066271 0.0011771091
#>                V19            V2           V20           V21          V22
#>  [1,] 0.0004528745  1.146907e-05  0.0002012671  3.538585e-04 0.0007672702
#>  [2,] 0.0006969145  5.417641e-05  0.0006299241  8.063525e-04 0.0010720489
#>  [3,] 0.0015360086  1.627362e-04  0.0017656938  2.057048e-03 0.0022011779
#>  [4,] 0.0019905252  2.270707e-04  0.0024744588  2.768122e-03 0.0026818207
#>  [5,] 0.0016390917  1.802700e-04  0.0020250574  2.301304e-03 0.0023050271
#>  [6,] 0.0001191137 -3.043770e-05 -0.0001760245 -9.358772e-05 0.0002859727
#>  [7,] 0.0009464975  8.346559e-05  0.0009384700  1.136156e-03 0.0013901295
#>  [8,] 0.0018612369  2.072271e-04  0.0023230262  2.627987e-03 0.0025942021
#>  [9,] 0.0029509828  3.562725e-04  0.0038376874  4.272072e-03 0.0040243641
#> [10,] 0.0014920596  1.722013e-04  0.0018399710  2.063230e-03 0.0020086010
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0012820330 0.0015680427 0.0016524092 0.0015886865 0.0015554042
#>  [2,] 0.0014379312 0.0017468534 0.0018931749 0.0019683549 0.0019908002
#>  [3,] 0.0024188561 0.0028133992 0.0030002658 0.0032908925 0.0034257438
#>  [4,] 0.0026380333 0.0029361495 0.0030883190 0.0034940431 0.0036471367
#>  [5,] 0.0022911594 0.0025986196 0.0027493055 0.0031087208 0.0033344017
#>  [6,] 0.0007119647 0.0008521309 0.0008132996 0.0007180139 0.0007514305
#>  [7,] 0.0017113436 0.0020035541 0.0021053808 0.0022116796 0.0022650530
#>  [8,] 0.0025640433 0.0028818394 0.0030435509 0.0034555409 0.0036775272
#>  [9,] 0.0037856717 0.0042031480 0.0044258220 0.0051296836 0.0054663397
#> [10,] 0.0019933462 0.0022289689 0.0023697808 0.0026801075 0.0027796006
#>                V28         V29           V3          V30         V31
#>  [1,] 0.0015248737 0.001470957 4.790525e-05 0.0015257096 0.001765720
#>  [2,] 0.0019634584 0.001881210 6.306921e-05 0.0017791099 0.001700129
#>  [3,] 0.0033752760 0.003149560 1.412997e-04 0.0027167687 0.002152752
#>  [4,] 0.0035962418 0.003330164 1.848668e-04 0.0028119153 0.002011289
#>  [5,] 0.0032464316 0.002917601 1.474096e-04 0.0022348230 0.001401536
#>  [6,] 0.0007558922 0.000736297 8.751503e-06 0.0007073326 0.001006907
#>  [7,] 0.0022377543 0.002113885 9.151132e-05 0.0019552692 0.001856222
#>  [8,] 0.0035982121 0.003270410 1.679833e-04 0.0025872305 0.001704899
#>  [9,] 0.0054090148 0.004976419 2.685911e-04 0.0040028479 0.002564978
#> [10,] 0.0027620157 0.002574766 1.422307e-04 0.0022216351 0.001663712
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0014188361 0.0013727515 0.0015148930 0.0017910731 0.0020758561
#>  [2,] 0.0014514469 0.0013546913 0.0013140651 0.0014651030 0.0016244195
#>  [3,] 0.0019795340 0.0017476754 0.0014021425 0.0014080184 0.0014999585
#>  [4,] 0.0019348741 0.0016962686 0.0012772709 0.0011778443 0.0011520218
#>  [5,] 0.0014124521 0.0011302793 0.0005414225 0.0003358080 0.0002967351
#>  [6,] 0.0007798335 0.0008110562 0.0009513771 0.0011242159 0.0013816490
#>  [7,] 0.0016104963 0.0014893805 0.0014285736 0.0015575136 0.0017477520
#>  [8,] 0.0016891513 0.0014088875 0.0008286522 0.0006328966 0.0006062571
#>  [9,] 0.0025963509 0.0021816057 0.0013249438 0.0009737107 0.0008219395
#> [10,] 0.0015806673 0.0014002509 0.0011265000 0.0010635242 0.0010411610
#>                V37          V38          V39            V4          V40
#>  [1,] 0.0017934449 0.0012230363 0.0007999955  1.476880e-05 0.0010239146
#>  [2,] 0.0014351506 0.0011390657 0.0008266818  7.209911e-05 0.0009145414
#>  [3,] 0.0014698009 0.0014763595 0.0012776997  2.185862e-04 0.0012488325
#>  [4,] 0.0012554633 0.0014924673 0.0014348641  3.063702e-04 0.0012523750
#>  [5,] 0.0004979103 0.0008775253 0.0009653145  2.549208e-04 0.0008213948
#>  [6,] 0.0011871332 0.0007506813 0.0005444164 -3.581463e-05 0.0007649220
#>  [7,] 0.0016142472 0.0013192122 0.0010270841  1.096828e-04 0.0011236068
#>  [8,] 0.0007798951 0.0011460623 0.0012041120  2.898762e-04 0.0010309325
#>  [9,] 0.0011125460 0.0018154936 0.0019468796  4.734492e-04 0.0015620985
#> [10,] 0.0011056367 0.0012311043 0.0011478449  2.267395e-04 0.0010182806
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0009127872 0.0004736516  1.148496e-04 -6.384916e-05 -0.0002702269
#>  [2,] 0.0009267775 0.0007582635  4.774366e-04  3.199460e-04  0.0002143370
#>  [3,] 0.0014086053 0.0015885949  1.322541e-03  1.141011e-03  0.0011476919
#>  [4,] 0.0014344084 0.0018426467  1.684103e-03  1.514104e-03  0.0016388027
#>  [5,] 0.0010621029 0.0015546553  1.506661e-03  1.417139e-03  0.0014892563
#>  [6,] 0.0006044368 0.0002300019 -3.697784e-05 -2.005494e-04 -0.0004121764
#>  [7,] 0.0011142803 0.0009665525  6.681448e-04  4.904189e-04  0.0003959434
#>  [8,] 0.0012545631 0.0017615047  1.673139e-03  1.551018e-03  0.0016437184
#>  [9,] 0.0019279901 0.0028267710  2.738605e-03  2.552005e-03  0.0028186901
#> [10,] 0.0011356624 0.0013846660  1.244456e-03  1.095592e-03  0.0011870597
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0001896380 -5.006888e-05 -5.998895e-05 -3.970991e-05  5.434265e-05
#>  [2,]  0.0001867274  1.502424e-04  1.138859e-04  5.699102e-05  1.432001e-04
#>  [3,]  0.0009293392  6.017186e-04  4.749697e-04  2.599213e-04  3.848918e-04
#>  [4,]  0.0013274892  8.097704e-04  6.470622e-04  3.614729e-04  4.976551e-04
#>  [5,]  0.0011807137  7.832858e-04  6.185442e-04  3.439396e-04  4.361908e-04
#>  [6,] -0.0003217196 -1.163026e-04 -1.141670e-04 -6.228257e-05 -1.094409e-05
#>  [7,]  0.0003389159  2.539149e-04  1.899436e-04  1.017526e-04  2.048298e-04
#>  [8,]  0.0013037094  8.363872e-04  6.697696e-04  3.728845e-04  4.882335e-04
#>  [9,]  0.0022528518  1.370418e-03  1.104535e-03  6.169099e-04  7.753947e-04
#> [10,]  0.0009782001  5.951921e-04  4.702396e-04  2.617512e-04  3.598463e-04
#>                 V50           V51           V52          V53          V54
#>  [1,]  1.614261e-05 -1.784959e-05 -1.056553e-05 1.186637e-05 1.288658e-05
#>  [2,]  3.756374e-05  1.226578e-05  1.507398e-05 1.759690e-05 1.382068e-05
#>  [3,]  9.077858e-05  7.513348e-05  6.576847e-05 3.721138e-05 2.553009e-05
#>  [4,]  1.177068e-04  1.124868e-04  9.537502e-05 4.981754e-05 3.584249e-05
#>  [5,]  9.481101e-05  9.991469e-05  8.119923e-05 3.248355e-05 1.929836e-05
#>  [6,] -2.968886e-06 -2.555162e-05 -2.230397e-05 4.293505e-06 5.889749e-06
#>  [7,]  5.055905e-05  2.767031e-05  2.494248e-05 2.449234e-05 1.998369e-05
#>  [8,]  1.082110e-04  1.107989e-04  9.163462e-05 4.101590e-05 2.651221e-05
#>  [9,]  1.822925e-04  1.921316e-04  1.594597e-04 7.265947e-05 4.904805e-05
#> [10,]  9.066855e-05  8.308847e-05  6.971623e-05 3.807330e-05 2.862166e-05
#>                V55          V56          V57           V58          V59
#>  [1,] 2.103519e-05 1.076664e-05 1.569301e-05 -2.723883e-06 1.612650e-05
#>  [2,] 2.041734e-05 1.342015e-05 1.349798e-05  3.739689e-06 1.824365e-05
#>  [3,] 2.752329e-05 2.445635e-05 1.817629e-05  1.944261e-05 3.128636e-05
#>  [4,] 3.191997e-05 3.164538e-05 2.031451e-05  3.427917e-05 3.742232e-05
#>  [5,] 1.625628e-05 2.098120e-05 1.375109e-05  2.493106e-05 2.777983e-05
#>  [6,] 1.143927e-05 3.568751e-06 8.374557e-06 -4.771235e-06 9.334505e-06
#>  [7,] 2.407354e-05 1.719358e-05 1.671296e-05  9.349608e-06 2.301027e-05
#>  [8,] 2.292506e-05 2.577064e-05 1.660670e-05  2.929724e-05 3.268784e-05
#>  [9,] 3.909624e-05 4.370781e-05 2.435657e-05  5.349174e-05 5.180066e-05
#> [10,] 2.626323e-05 2.541392e-05 1.690000e-05  2.666238e-05 2.877031e-05
#>                 V6          V60           V7            V8            V9
#>  [1,] 1.710152e-04 1.808729e-05 1.976723e-04  1.715565e-05 -2.849748e-04
#>  [2,] 2.109316e-04 1.828012e-05 2.588794e-04  1.824595e-04  6.966686e-05
#>  [3,] 4.150490e-04 2.620627e-05 5.130232e-04  5.900479e-04  7.613950e-04
#>  [4,] 5.056874e-04 2.839057e-05 6.149731e-04  8.275537e-04  1.226122e-03
#>  [5,] 4.041655e-04 1.909165e-05 5.074386e-04  7.036197e-04  1.080964e-03
#>  [6,] 7.035327e-05 9.907356e-06 7.014211e-05 -1.190033e-04 -4.485335e-04
#>  [7,] 2.867627e-04 2.112047e-05 3.376046e-04  2.901188e-04  2.240074e-04
#>  [8,] 4.614378e-04 2.334040e-05 5.715718e-04  7.890298e-04  1.200979e-03
#>  [9,] 7.315654e-04 3.638231e-05 9.011646e-04  1.319304e-03  2.097071e-03
#> [10,] 3.891286e-04 2.291148e-05 4.698792e-04  6.241367e-04  9.067493e-04
#> 
#> $vW[[2]]
#>             [,1]       [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] -0.03619808 -0.0170488 -0.02703978 -0.0484864 -0.01289887 -0.05845462
#> [2,]  0.03619808  0.0170488  0.02703978  0.0484864  0.01289887  0.05845462
#>            [,7]        [,8]        [,9]      [,10]
#> [1,] -0.0408077 -0.01921122 -0.03130711 -0.0559584
#> [2,]  0.0408077  0.01921122  0.03130711  0.0559584
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4381951 0.1319771 0.3153886 0.5089543 0.2235086 0.7085576 0.5503210
#>  [8] 0.3785386 0.2504846 0.4387658
#> 
#> $B[[2]]
#> [1]  0.01096881 -0.08751854
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.002516144 0.002762884 0.004244922 0.004519967 0.003562233 0.001394328
#>  [7] 0.003170630 0.004135376 0.006250149 0.003530823
#> 
#> $vB[[2]]
#> [1] -0.0712592  0.0712592
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#>  [2,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#>  [3,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#>  [4,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#>  [5,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#>  [6,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#>  [7,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#>  [8,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#>  [9,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [10,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [11,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [12,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [13,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [14,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [15,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [16,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [17,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [18,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [19,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [20,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [21,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [22,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [23,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [24,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [25,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [26,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [27,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [28,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [29,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [30,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [31,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [32,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [33,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [34,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [35,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [36,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [37,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [38,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [39,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [40,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#>  [2,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#>  [3,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#>  [4,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#>  [5,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#>  [6,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#>  [7,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#>  [8,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#>  [9,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [10,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [11,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [12,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [13,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [14,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [15,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [16,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [17,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [18,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [19,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [20,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [21,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [22,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [23,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [24,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [25,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [26,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [27,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [28,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [29,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [30,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [31,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [32,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [33,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [34,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [35,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [36,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [37,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [38,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [39,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [40,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#>  [2,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#>  [3,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#>  [4,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#>  [5,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#>  [6,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#>  [7,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#>  [8,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#>  [9,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [10,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [11,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [12,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [13,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [14,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [15,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [16,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [17,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [18,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [19,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [20,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [21,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [22,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [23,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [24,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [25,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [26,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [27,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [28,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [29,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [30,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [31,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [32,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [33,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [34,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [35,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [36,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [37,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [38,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [39,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [40,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#>  [2,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#>  [3,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#>  [4,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#>  [5,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#>  [6,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#>  [7,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#>  [8,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#>  [9,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [10,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [11,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [12,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [13,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [14,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [15,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [16,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [17,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [18,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [19,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [20,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [21,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [22,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [23,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [24,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [25,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [26,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [27,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [28,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [29,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [30,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [31,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [32,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [33,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [34,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [35,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [36,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [37,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [38,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [39,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [40,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#>  [2,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#>  [3,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#>  [4,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#>  [5,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#>  [6,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#>  [7,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#>  [8,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#>  [9,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [10,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [11,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [12,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [13,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [14,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [15,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [16,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [17,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [18,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [19,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [20,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [21,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [22,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [23,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [24,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [25,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [26,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [27,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [28,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [29,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [30,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [31,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [32,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [33,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [34,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [35,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [36,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [37,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [38,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [39,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [40,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#>  [2,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#>  [3,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#>  [4,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#>  [5,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#>  [6,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#>  [7,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#>  [8,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#>  [9,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [10,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [11,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [12,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [13,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [14,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [15,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [16,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [17,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [18,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [19,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [20,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [21,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [22,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [23,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [24,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [25,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [26,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [27,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [28,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [29,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [30,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [31,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [32,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [33,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [34,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [35,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [36,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [37,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [38,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [39,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [40,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5905092 0.4070475 0.4714528 0.7041909 0.3073245 0.8602363 0.6804817
#>  [2,] 0.5833004 0.4762204 0.5749931 0.6931687 0.3943051 0.7642984 0.6254533
#>  [3,] 0.5910086 0.4401487 0.5923814 0.7365880 0.3575191 0.8234224 0.6723270
#>  [4,] 0.7529265 0.4590722 0.5786616 0.7942873 0.3859280 0.9381072 0.8135591
#>  [5,] 0.7446425 0.5103514 0.6290133 0.7909375 0.4655126 0.9130819 0.7994719
#>  [6,] 0.4795426 0.3274946 0.5241346 0.6981564 0.2333892 0.8049182 0.5845881
#>  [7,] 0.4127596 0.2785103 0.4079178 0.6140545 0.1915441 0.7287274 0.4393864
#>  [8,] 0.6849512 0.3807372 0.5181462 0.7352155 0.3090277 0.9180472 0.7409526
#>  [9,] 0.7523090 0.5610259 0.6745072 0.8081459 0.5066495 0.9174537 0.8115013
#> [10,] 0.6285143 0.3935245 0.5434593 0.7631468 0.2895692 0.8823778 0.7020430
#> [11,] 0.4599973 0.3933884 0.5565635 0.6710064 0.3121794 0.7115151 0.5425469
#> [12,] 0.7312268 0.5289805 0.6299399 0.7814446 0.4662715 0.9021929 0.7822256
#> [13,] 0.6430610 0.3625343 0.5269616 0.7449214 0.3125035 0.8950718 0.7273942
#> [14,] 0.4964591 0.3945740 0.5461436 0.7148995 0.3005982 0.7863908 0.5785130
#> [15,] 0.4783870 0.3302091 0.5063635 0.6679045 0.2457588 0.7867149 0.5512857
#> [16,] 0.5970317 0.4229781 0.5503570 0.7362133 0.3679304 0.8497619 0.7111980
#> [17,] 0.6630421 0.4525650 0.5853473 0.7641216 0.3948520 0.8846304 0.7472333
#> [18,] 0.6526426 0.4055329 0.5406898 0.7738392 0.3055748 0.8914505 0.7262354
#> [19,] 0.6724730 0.4431726 0.5210162 0.7431414 0.4123157 0.8738875 0.7381991
#> [20,] 0.7108589 0.5621826 0.6423465 0.7990418 0.4796421 0.8938882 0.7824691
#> [21,] 0.6615833 0.5101197 0.6564693 0.7692927 0.4223537 0.8741018 0.7296571
#> [22,] 0.5454416 0.2879465 0.3612334 0.6220283 0.2282107 0.8504767 0.6193522
#> [23,] 0.6696842 0.4940859 0.6268805 0.7727935 0.4393063 0.8700887 0.7468663
#> [24,] 0.6395681 0.4183992 0.5707463 0.7801000 0.3345521 0.9058221 0.7161491
#> [25,] 0.6316519 0.4553036 0.6023748 0.7684508 0.3634497 0.8888953 0.7104233
#> [26,] 0.6862358 0.4425566 0.6018133 0.7703617 0.4105388 0.8974583 0.7546678
#> [27,] 0.6974295 0.5210288 0.6000248 0.7812650 0.4496414 0.8898685 0.7667705
#> [28,] 0.4591331 0.3763029 0.5430796 0.7109434 0.2885516 0.7612796 0.5573718
#> [29,] 0.7331386 0.4936856 0.6264025 0.7893988 0.4437430 0.9118459 0.7949180
#> [30,] 0.6455344 0.4266831 0.5376750 0.7325679 0.3314068 0.8908949 0.7065450
#> [31,] 0.7255238 0.4150079 0.5639162 0.7619936 0.3712894 0.9268188 0.7953176
#> [32,] 0.7521676 0.5077348 0.6350768 0.7942018 0.4661236 0.9184704 0.8065517
#> [33,] 0.4066840 0.2940052 0.4156242 0.6217531 0.2011015 0.7194655 0.4260759
#> [34,] 0.7495867 0.5578690 0.6763251 0.8159102 0.4948872 0.9225422 0.8109985
#> [35,] 0.6665976 0.5054447 0.6883108 0.7924912 0.4236254 0.8816872 0.7471865
#> [36,] 0.4905323 0.3048374 0.3995674 0.6756306 0.1996442 0.8471438 0.5690033
#> [37,] 0.6651926 0.4203139 0.5690040 0.7365592 0.3571050 0.8814813 0.7053467
#> [38,] 0.7743465 0.4671079 0.5841891 0.7895440 0.4144013 0.9500656 0.8297092
#> [39,] 0.6567915 0.3582705 0.5011507 0.7210151 0.3123681 0.9014592 0.7151008
#> [40,] 0.6220103 0.4368132 0.5915513 0.7522599 0.3491598 0.8595260 0.6877589
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3698954 0.5790786 0.8198553
#>  [2,] 0.4860204 0.6381318 0.7471508
#>  [3,] 0.4519615 0.5931761 0.7930810
#>  [4,] 0.4605686 0.7001291 0.8949620
#>  [5,] 0.5193784 0.6974980 0.8599202
#>  [6,] 0.3117365 0.5279456 0.7888032
#>  [7,] 0.3101259 0.4677289 0.7279435
#>  [8,] 0.4441848 0.6109646 0.8739767
#>  [9,] 0.5423268 0.7480229 0.8751457
#> [10,] 0.3825739 0.6115008 0.8521406
#> [11,] 0.3929416 0.5815658 0.7207723
#> [12,] 0.5319101 0.7152461 0.8457577
#> [13,] 0.3953072 0.6202471 0.8483944
#> [14,] 0.3859592 0.5869024 0.8118165
#> [15,] 0.3172702 0.5188832 0.7755599
#> [16,] 0.4356492 0.5973658 0.8225329
#> [17,] 0.4587514 0.6546046 0.8459762
#> [18,] 0.4054447 0.6240751 0.8510250
#> [19,] 0.4844722 0.6329348 0.8233794
#> [20,] 0.5264989 0.6880583 0.8584994
#> [21,] 0.5119217 0.6839005 0.8456828
#> [22,] 0.3406785 0.4301088 0.7728051
#> [23,] 0.4990034 0.6641329 0.8330880
#> [24,] 0.4224335 0.6756492 0.8839845
#> [25,] 0.4307396 0.6819259 0.8751191
#> [26,] 0.4417534 0.6927183 0.8577602
#> [27,] 0.4917334 0.6561345 0.8456993
#> [28,] 0.3516540 0.5696245 0.7958975
#> [29,] 0.5051171 0.6861759 0.8650939
#> [30,] 0.4114412 0.6664208 0.8426669
#> [31,] 0.4391418 0.6895327 0.8761611
#> [32,] 0.5185782 0.7065573 0.8684876
#> [33,] 0.2885205 0.4725385 0.7317849
#> [34,] 0.5321246 0.7501557 0.8823308
#> [35,] 0.5078093 0.7134932 0.8572848
#> [36,] 0.3200911 0.5153346 0.8087054
#> [37,] 0.4486475 0.6642907 0.8421018
#> [38,] 0.4933573 0.6977483 0.9037567
#> [39,] 0.3829288 0.6198095 0.8531168
#> [40,] 0.4689002 0.6384192 0.8240150
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4186984 0.5813016
#>  [2,] 0.4096515 0.5903485
#>  [3,] 0.4119028 0.5880972
#>  [4,] 0.4057996 0.5942004
#>  [5,] 0.4018246 0.5981754
#>  [6,] 0.4229236 0.5770764
#>  [7,] 0.4307089 0.5692911
#>  [8,] 0.4143490 0.5856510
#>  [9,] 0.3966116 0.6033884
#> [10,] 0.4148046 0.5851954
#> [11,] 0.4165254 0.5834746
#> [12,] 0.4010331 0.5989669
#> [13,] 0.4146603 0.5853397
#> [14,] 0.4152245 0.5847755
#> [15,] 0.4239514 0.5760486
#> [16,] 0.4123016 0.5876984
#> [17,] 0.4078113 0.5921887
#> [18,] 0.4133348 0.5866652
#> [19,] 0.4095954 0.5904046
#> [20,] 0.4006052 0.5993948
#> [21,] 0.4030322 0.5969678
#> [22,] 0.4325812 0.5674188
#> [23,] 0.4043562 0.5956438
#> [24,] 0.4087918 0.5912082
#> [25,] 0.4071230 0.5928770
#> [26,] 0.4059858 0.5940142
#> [27,] 0.4049337 0.5950663
#> [28,] 0.4170138 0.5829862
#> [29,] 0.4031236 0.5968764
#> [30,] 0.4119795 0.5880205
#> [31,] 0.4083821 0.5916179
#> [32,] 0.4012286 0.5987714
#> [33,] 0.4301809 0.5698191
#> [34,] 0.3967779 0.6032221
#> [35,] 0.4006895 0.5993105
#> [36,] 0.4271719 0.5728281
#> [37,] 0.4098846 0.5901154
#> [38,] 0.4044579 0.5955421
#> [39,] 0.4161615 0.5838385
#> [40,] 0.4095979 0.5904021
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]         [,3]      [,4]        [,5]      [,6]
#>  [1,]  0.36607060 -0.37618461 -0.114313218 0.8673353 -0.81265747 1.8172540
#>  [2,]  0.33633694 -0.09519008  0.302252828 0.8149751 -0.42925163 1.1763919
#>  [3,]  0.36813639 -0.24055862  0.373818762 1.0283096 -0.58614832 1.5397091
#>  [4,]  1.11428160 -0.16407823  0.317281378 1.3509644 -0.46446154 2.7184596
#>  [5,]  1.07023948  0.04141142  0.527986314 1.3305855 -0.13816901 2.3518590
#>  [6,] -0.08187530 -0.71953868  0.096613325 0.8385344 -1.18927174 1.4173219
#>  [7,] -0.35256887 -0.95186293 -0.372579956 0.4643874 -1.44000837 0.9881755
#>  [8,]  0.77662007 -0.48642051  0.072616533 1.0212476 -0.80466880 2.4161047
#>  [9,]  1.11096507  0.24532674  0.728642223 1.4380074  0.02659960 2.4082425
#> [10,]  0.52584844 -0.43252110  0.174276894 1.1700097 -0.89747712 2.0151423
#> [11,] -0.16035342 -0.43309122  0.227226813 0.7127403 -0.78995011 0.9027537
#> [12,]  1.00085559  0.11605208  0.531958916 1.2741048 -0.13511935 2.2218314
#> [13,]  0.58867452 -0.56438121  0.107950946 1.0717068 -0.78844096 2.1436276
#> [14,] -0.01416394 -0.42812551  0.185101125 0.9193003 -0.84445105 1.3033059
#> [15,] -0.08650597 -0.70723960  0.025455293 0.6987224 -1.12136158 1.3052360
#> [16,]  0.39311212 -0.31055998  0.202113247 1.0263792 -0.54110592 1.7327345
#> [17,]  0.67688053 -0.19031241  0.344764248 1.1754107 -0.42696190 2.0370292
#> [18,]  0.63067535 -0.38246358  0.163119989 1.2301178 -0.82088968 2.1056436
#> [19,]  0.71939159 -0.22829585  0.084114383 1.0623606 -0.35440045 1.9357769
#> [20,]  0.89955907  0.25002492  0.585563166 1.3803165 -0.08147672 2.1310877
#> [21,]  0.67035772  0.04048426  0.647599440 1.2043219 -0.31311872 1.9377233
#> [22,]  0.18226928 -0.90537857 -0.570014980 0.4981661 -1.21844243 1.7383449
#> [23,]  0.70675718 -0.02365753  0.518857276 1.2241523 -0.24397808 1.9017428
#> [24,]  0.57348988 -0.32934824  0.284896750 1.2662492 -0.68766775 2.2636577
#> [25,]  0.53930997 -0.17926417  0.415369875 1.1995843 -0.56042255 2.0795063
#> [26,]  0.78257977 -0.23079253  0.413026129 1.2103548 -0.36173857 2.1692971
#> [27,]  0.83508689  0.08416467  0.405568495 1.2730534 -0.20211954 2.0893985
#> [28,] -0.16383292 -0.50527033  0.172746844 0.8999702 -0.90242904 1.1597078
#> [29,]  1.01060552 -0.02525893  0.516814168 1.3213053 -0.22598467 2.3363851
#> [30,]  0.59946723 -0.29539697  0.150986197 1.0076905 -0.70182938 2.0999151
#> [31,]  0.97202921 -0.34330090  0.257071291 1.1636404 -0.52668905 2.5388198
#> [32,]  1.11020626  0.03094160  0.554059083 1.3504414 -0.13571356 2.4217430
#> [33,] -0.37769073 -0.87601043 -0.340762665 0.4969958 -1.37942434 0.9418120
#> [34,]  1.09640915  0.23251810  0.736934253 1.4888805 -0.02045195 2.4774004
#> [35,]  0.69283660  0.02177969  0.792233829 1.3400075 -0.30790819 2.0085057
#> [36,] -0.03787513 -0.82436703 -0.407267830 0.7337634 -1.38851972 1.7123731
#> [37,]  0.68652102 -0.32148510  0.277788825 1.0281612 -0.58795160 2.0065328
#> [38,]  1.23301884 -0.13175883  0.339994020 1.3221792 -0.34579997 2.9458205
#> [39,]  0.64902865 -0.58287852  0.004602618 0.9495022 -0.78907130 2.2135439
#> [40,]  0.49808966 -0.25410577  0.370381976 1.1107015 -0.62273451 1.8113590
#>             [,7]         [,8]        [,9]     [,10]
#>  [1,]  0.7559864 -0.532665583  0.31899211 1.5153675
#>  [2,]  0.5127601 -0.055933014  0.56726462 1.0834737
#>  [3,]  0.7187288 -0.192748386  0.37711103 1.3435981
#>  [4,]  1.4733040 -0.158053950  0.84791252 2.1424588
#>  [5,]  1.3829969  0.077552517  0.83541158 1.8146276
#>  [6,]  0.3416371 -0.792013672  0.11189902 1.3177265
#>  [7,] -0.2436526 -0.799530985 -0.12926417 0.9842137
#>  [8,]  1.0509258 -0.224195164  0.45136865 1.9365871
#>  [9,]  1.4597950  0.169713518  1.08809561 1.9472430
#> [10,]  0.8570454 -0.478637376  0.45362518 1.7514896
#> [11,]  0.1706003 -0.434964187  0.32920435 0.9482955
#> [12,]  1.2786834  0.127814179  0.92100131 1.7017082
#> [13,]  0.9814417 -0.425057550  0.49059730 1.7220634
#> [14,]  0.3166722 -0.464329882  0.35117482 1.4618566
#> [15,]  0.2058670 -0.766345284  0.07556873 1.2399762
#> [16,]  0.9012098 -0.258838817  0.39450134 1.5336031
#> [17,]  1.0839105 -0.165370408  0.63934149 1.7033838
#> [18,]  0.9756054 -0.382829221  0.50688125 1.7426631
#> [19,]  1.0366293 -0.062131296  0.54482784 1.5394133
#> [20,]  1.2801137  0.106095180  0.79105734 1.8028821
#> [21,]  0.9928834  0.047695741  0.77175530 1.7011341
#> [22,]  0.4867994 -0.660272089 -0.28140735 1.2242186
#> [23,]  1.0819685 -0.003986392  0.68176676 1.6076724
#> [24,]  0.9254395 -0.312791532  0.73384818 2.0307157
#> [25,]  0.8974408 -0.278834139  0.76263667 1.9469998
#> [26,]  1.1236644 -0.234049011  0.81285841 1.7968104
#> [27,]  1.1901648 -0.033069321  0.64611515 1.7012608
#> [28,]  0.2305024 -0.611776618  0.28031926 1.3608483
#> [29,]  1.3548293  0.020469079  0.78230137 1.8582588
#> [30,]  0.8786626 -0.358010776  0.69204094 1.6782068
#> [31,]  1.3572822 -0.244645757  0.79793544 1.9565681
#> [32,]  1.4277577  0.074346940  0.87872186 1.8876524
#> [33,] -0.2978795 -0.902580497 -0.10995648 1.0036971
#> [34,]  1.4565113  0.128675696  1.09944284 2.0146901
#> [35,]  1.0836629  0.031239674  0.91241048 1.7929193
#> [36,]  0.2777857 -0.753353204  0.06135778 1.4416199
#> [37,]  0.8728901 -0.206136993  0.68247434 1.6739506
#> [38,]  1.5835677 -0.026572464  0.83659843 2.2396807
#> [39,]  0.9202883 -0.477135331  0.48873962 1.7592584
#> [40,]  0.7896629 -0.124559944  0.56850977 1.5437900
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.1661362 0.1619826
#>  [2,] -0.1754239 0.1899824
#>  [3,] -0.1714590 0.1846460
#>  [4,] -0.1906226 0.1907347
#>  [5,] -0.1909887 0.2068798
#>  [6,] -0.1588702 0.1519128
#>  [7,] -0.1424645 0.1364950
#>  [8,] -0.1732606 0.1727549
#>  [9,] -0.1994288 0.2201747
#> [10,] -0.1740928 0.1700453
#> [11,] -0.1659450 0.1711085
#> [12,] -0.1919151 0.2092472
#> [13,] -0.1756893 0.1690433
#> [14,] -0.1676682 0.1747403
#> [15,] -0.1562642 0.1503088
#> [16,] -0.1726110 0.1818478
#> [17,] -0.1823866 0.1906343
#> [18,] -0.1761436 0.1740527
#> [19,] -0.1770927 0.1885456
#> [20,] -0.1892016 0.2137425
#> [21,] -0.1866609 0.2061859
#> [22,] -0.1397442 0.1315833
#> [23,] -0.1847781 0.2025684
#> [24,] -0.1836879 0.1852746
#> [25,] -0.1853060 0.1905654
#> [26,] -0.1892966 0.1912886
#> [27,] -0.1834940 0.2014555
#> [28,] -0.1655385 0.1695055
#> [29,] -0.1893801 0.2030869
#> [30,] -0.1801576 0.1756306
#> [31,] -0.1882882 0.1823697
#> [32,] -0.1928154 0.2075331
#> [33,] -0.1438988 0.1372144
#> [34,] -0.1997889 0.2191200
#> [35,] -0.1926271 0.2099659
#> [36,] -0.1509398 0.1424592
#> [37,] -0.1814711 0.1829713
#> [38,] -0.1907537 0.1961709
#> [39,] -0.1745769 0.1639739
#> [40,] -0.1772481 0.1883799
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4186984  0.4186984
#>  [2,] -0.4096515  0.4096515
#>  [3,]  0.5880972 -0.5880972
#>  [4,]  0.5942004 -0.5942004
#>  [5,]  0.5981754 -0.5981754
#>  [6,]  0.5770764 -0.5770764
#>  [7,]  0.5692911 -0.5692911
#>  [8,] -0.4143490  0.4143490
#>  [9,] -0.3966116  0.3966116
#> [10,]  0.5851954 -0.5851954
#> [11,] -0.4165254  0.4165254
#> [12,] -0.4010331  0.4010331
#> [13,] -0.4146603  0.4146603
#> [14,]  0.5847755 -0.5847755
#> [15,]  0.5760486 -0.5760486
#> [16,] -0.4123016  0.4123016
#> [17,]  0.5921887 -0.5921887
#> [18,]  0.5866652 -0.5866652
#> [19,]  0.5904046 -0.5904046
#> [20,]  0.5993948 -0.5993948
#> [21,] -0.4030322  0.4030322
#> [22,]  0.5674188 -0.5674188
#> [23,]  0.5956438 -0.5956438
#> [24,] -0.4087918  0.4087918
#> [25,] -0.4071230  0.4071230
#> [26,]  0.5940142 -0.5940142
#> [27,]  0.5950663 -0.5950663
#> [28,]  0.5829862 -0.5829862
#> [29,]  0.5968764 -0.5968764
#> [30,] -0.4119795  0.4119795
#> [31,]  0.5916179 -0.5916179
#> [32,]  0.5987714 -0.5987714
#> [33,]  0.5698191 -0.5698191
#> [34,] -0.3967779  0.3967779
#> [35,] -0.4006895  0.4006895
#> [36,] -0.4271719  0.4271719
#> [37,] -0.4098846  0.4098846
#> [38,] -0.4044579  0.4044579
#> [39,]  0.5838385 -0.5838385
#> [40,]  0.5904021 -0.5904021
#> 
#> $L
#> [1] 0.6977327 0.7255278 0.7174145 0.6921490 0.8015418 0.7326427
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4637681