Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11         V12          V13
#>  [1,] -0.4165935 -0.150947808 -0.036882491 -0.01238269 -0.056694287
#>  [2,] -0.2731001 -0.144604048 -0.103924586 -0.01545742 -0.109974121
#>  [3,] -0.3142787 -0.229923212 -0.191891099 -0.09131284 -0.062258286
#>  [4,] -0.3678008 -0.058612629 -0.079067312 -0.10518677 -0.093226663
#>  [5,] -0.1999668 -0.053167997 -0.065745341 -0.16533320 -0.110002105
#>  [6,] -0.3575150 -0.027209202 -0.080831027 -0.12394379 -0.045986599
#>  [7,] -0.2793690 -0.064901641 -0.164663112 -0.19365051 -0.183083855
#>  [8,] -0.3079411 -0.132961310 -0.111959942 -0.02507124 -0.007511397
#>  [9,] -0.2819142 -0.118244778  0.007901320 -0.10610532 -0.019757663
#> [10,] -0.2713950 -0.001956438  0.007042539 -0.06457549 -0.016188050
#>                V14          V15         V16         V17         V18         V19
#>  [1,] -0.123708661 -0.098593872  0.05874828  0.02488069  0.13022035  0.10630536
#>  [2,] -0.030220197 -0.061363927 -0.06447385 -0.03411202 -0.12056260 -0.06370721
#>  [3,] -0.041688362 -0.009896309 -0.15663903 -0.14220925 -0.12027030  0.04495593
#>  [4,] -0.033922499 -0.052147336 -0.07482085 -0.01424370  0.12794739  0.03628462
#>  [5,] -0.085727219 -0.095680005 -0.07692542  0.00950168 -0.05434841 -0.02008568
#>  [6,] -0.036212263 -0.085329594  0.06135504 -0.01895087  0.05144582  0.09603124
#>  [7,] -0.003777837 -0.099223989  0.04132967  0.12173461  0.08781162  0.19702318
#>  [8,] -0.166676679  0.010081123 -0.01971866 -0.01069117  0.04758767  0.07676979
#>  [9,] -0.067575687 -0.020670098  0.11040573  0.06092280  0.15773814  0.05919137
#> [10,] -0.080160585 -0.040863546 -0.08568912  0.09448374  0.09095466 -0.03787536
#>               V2           V20           V21        V22         V23        V24
#>  [1,] -0.3044911  0.2355243420  0.1577759094 0.14531215  0.25986553 0.28334654
#>  [2,] -0.3190957  0.0006822193  0.1538282975 0.02771210  0.11288992 0.07524394
#>  [3,] -0.2066771  0.0111213943 -0.0208367540 0.08562822 -0.04843111 0.17447932
#>  [4,] -0.3177629  0.1417502835  0.2128694260 0.24647316  0.15408062 0.23725423
#>  [5,] -0.2575030 -0.0463332758  0.0483059787 0.07177723  0.12147667 0.13803661
#>  [6,] -0.2193187 -0.0142389015  0.0368420085 0.12477186  0.17223422 0.17027882
#>  [7,] -0.3297510 -0.0018467609 -0.0068472497 0.04068915  0.05829211 0.01610663
#>  [8,] -0.2549746 -0.0277788753  0.0594351145 0.06457122 -0.04565203 0.13133219
#>  [9,] -0.2263086  0.2075933353  0.0942597230 0.16137062  0.19451946 0.07202698
#> [10,] -0.2287957  0.0566954061  0.0007527496 0.01729429 -0.02135087 0.13373602
#>              V25        V26        V27        V28         V29         V3
#>  [1,] 0.22141079 0.24924520 0.24715800 0.32500341  0.29908889 -0.2936533
#>  [2,] 0.09093585 0.06382349 0.10944518 0.19144667  0.12235589 -0.3538508
#>  [3,] 0.08405629 0.03985852 0.01359891 0.02757329  0.10633894 -0.1925204
#>  [4,] 0.17714952 0.17274508 0.31890491 0.14092682  0.12899192 -0.2518370
#>  [5,] 0.14274018 0.08703597 0.22981980 0.15880829  0.04090313 -0.2736134
#>  [6,] 0.08993003 0.27245365 0.30191752 0.28096853  0.26703005 -0.3434745
#>  [7,] 0.12482701 0.08329870 0.08380125 0.02352292 -0.07367845 -0.2459310
#>  [8,] 0.19944718 0.05247450 0.04768986 0.13776012  0.08641317 -0.2248564
#>  [9,] 0.19134811 0.23986407 0.10142208 0.08383472  0.01504277 -0.1351889
#> [10,] 0.22401625 0.14589565 0.19496505 0.17358573  0.10596988 -0.3619972
#>               V30         V31         V32         V33          V34
#>  [1,]  0.28662632  0.10313285  0.05918371  0.07580660  0.107244915
#>  [2,] -0.02235761  0.10036150 -0.08430279 -0.04097197  0.029598492
#>  [3,]  0.07846968 -0.07153060 -0.01164918 -0.01076783 -0.037569008
#>  [4,]  0.02378594  0.02656591  0.03879397  0.02285830  0.018096595
#>  [5,]  0.14497735  0.09333438 -0.01976025 -0.10047112 -0.071790336
#>  [6,]  0.15365583  0.10427525  0.08355599  0.07035508 -0.007706400
#>  [7,] -0.10580752 -0.08779083 -0.03097534 -0.07940048  0.002670399
#>  [8,]  0.04300722 -0.03638855 -0.02172795 -0.04313142 -0.057990714
#>  [9,] -0.01879304 -0.06014886 -0.02192593 -0.00326985  0.025560873
#> [10,]  0.19897731  0.11217745  0.01127543  0.08752101 -0.042601875
#>                 V35           V36         V37         V38         V39
#>  [1,]  0.0096271372  0.0199201747  0.12843594  0.01534277 -0.06846865
#>  [2,] -0.0374833094  0.0605844008 -0.12135429 -0.06531120 -0.11214500
#>  [3,] -0.0442359118 -0.0754100115 -0.12227796 -0.17486245 -0.21628190
#>  [4,]  0.1347316831 -0.0001463561 -0.03732149 -0.03298545 -0.02856585
#>  [5,] -0.1045701785 -0.0981709041 -0.12088511 -0.11263122 -0.07445125
#>  [6,] -0.0489709590 -0.0656348161  0.04834854 -0.01493401 -0.11930384
#>  [7,]  0.0071078812 -0.1120056385 -0.10293873 -0.14014657 -0.17217943
#>  [8,]  0.0007558604 -0.0404871503  0.02624807 -0.10311043  0.02781117
#>  [9,] -0.0047820974 -0.0832011852  0.03421265  0.01079119  0.04078634
#> [10,]  0.0176909788  0.0647322682  0.02548510  0.06748874  0.03156223
#>               V4          V40         V41         V42          V43          V44
#>  [1,] -0.2825423 -0.047167103  0.06009247 -0.05016202 -0.098005255 -0.146502006
#>  [2,] -0.2156642  0.001150236 -0.07105987 -0.18147151 -0.165771406 -0.065631196
#>  [3,] -0.2255137 -0.195227074 -0.03936273 -0.16289433 -0.024558877 -0.156728092
#>  [4,] -0.3427663 -0.014799218 -0.07622544  0.04471619  0.001156794 -0.155719697
#>  [5,] -0.2813412 -0.165554854 -0.02112508 -0.06672630 -0.150891331 -0.060389039
#>  [6,] -0.2540864 -0.124997428 -0.14161993 -0.08000777 -0.020626472 -0.021788783
#>  [7,] -0.1907295 -0.144263020 -0.16814508 -0.03972743 -0.194528088 -0.173079012
#>  [8,] -0.3195191  0.030672791 -0.04859808 -0.02343702 -0.066135021 -0.078698050
#>  [9,] -0.1989433 -0.077848272 -0.05741061 -0.03432337 -0.124311386 -0.004073326
#> [10,] -0.2531141  0.006145386 -0.06516115 -0.11307745 -0.012674789 -0.025108226
#>                V45         V46         V47        V48        V49         V5
#>  [1,] -0.047189338 -0.06193881 -0.17951411 -0.2172475 -0.3606978 -0.1655846
#>  [2,] -0.152267882 -0.08438603 -0.13127505 -0.2446912 -0.3318410 -0.2200371
#>  [3,] -0.125503019 -0.10329208 -0.22062958 -0.1748021 -0.2130376 -0.3216072
#>  [4,] -0.121885937 -0.17248590 -0.24521599 -0.2484470 -0.3521724 -0.3081074
#>  [5,] -0.144406336 -0.23877051 -0.09694225 -0.1306381 -0.1904006 -0.2158386
#>  [6,] -0.203558460 -0.12557199 -0.21129452 -0.2719483 -0.3051774 -0.1747029
#>  [7,] -0.156965654 -0.10675497 -0.25209785 -0.1710315 -0.1908835 -0.1558187
#>  [8,] -0.071258849 -0.08562540 -0.16977076 -0.1431143 -0.2658825 -0.1542931
#>  [9,] -0.096601886 -0.16445383 -0.18510194 -0.1430288 -0.2538773 -0.1296920
#> [10,] -0.008945413 -0.19556731 -0.18164290 -0.2156607 -0.2701731 -0.3135858
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3749483 -0.3482950 -0.3531042 -0.3386404 -0.4467694 -0.3956269
#>  [2,] -0.3108718 -0.3194924 -0.4170846 -0.3448482 -0.3228357 -0.2997478
#>  [3,] -0.2485615 -0.2055660 -0.3021735 -0.2865735 -0.2257295 -0.3217074
#>  [4,] -0.2612253 -0.2909904 -0.3327127 -0.3448064 -0.3777457 -0.3412175
#>  [5,] -0.3108972 -0.2723467 -0.3052698 -0.3781345 -0.3145156 -0.2649584
#>  [6,] -0.2685545 -0.2582907 -0.2210190 -0.3784810 -0.2749445 -0.3931472
#>  [7,] -0.3324473 -0.3068035 -0.2263540 -0.2641062 -0.4064727 -0.2481784
#>  [8,] -0.3025859 -0.4297881 -0.4025689 -0.3242503 -0.3480765 -0.4134168
#>  [9,] -0.2374633 -0.2143964 -0.1764964 -0.2126713 -0.2904037 -0.2958059
#> [10,] -0.3839794 -0.4121168 -0.3265709 -0.4358402 -0.2783674 -0.4016326
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.4255869 -0.4651209 -0.3968839 -0.3585833 -0.26741915 -0.3742060
#>  [2,] -0.4017659 -0.3050612 -0.2816566 -0.3395956 -0.21704605 -0.2900180
#>  [3,] -0.1983968 -0.2979757 -0.2578574 -0.3604659 -0.16420975 -0.2015261
#>  [4,] -0.4235686 -0.2709150 -0.3072880 -0.3611926 -0.20898321 -0.3807877
#>  [5,] -0.2754907 -0.4155853 -0.2655481 -0.4470200 -0.19846227 -0.4483493
#>  [6,] -0.2578314 -0.2770318 -0.3547852 -0.2726032 -0.25138324 -0.3379069
#>  [7,] -0.4139441 -0.4033262 -0.2587646 -0.2828111 -0.26430872 -0.2620978
#>  [8,] -0.2643169 -0.2518949 -0.4067682 -0.3606542 -0.27134686 -0.3284085
#>  [9,] -0.2123056 -0.2676567 -0.2387419 -0.1949566 -0.16479868 -0.2530297
#> [10,] -0.3105824 -0.3660074 -0.4246525 -0.3735867 -0.08764479 -0.4009801
#>                V7         V8          V9
#>  [1,] -0.25241197 -0.2336221 -0.09256853
#>  [2,] -0.17174118 -0.2209451 -0.10277276
#>  [3,] -0.11439605 -0.1285072 -0.20418425
#>  [4,] -0.18899401 -0.1432126 -0.10279866
#>  [5,] -0.20027446 -0.1589081 -0.16476078
#>  [6,] -0.23723393 -0.1079231 -0.15945491
#>  [7,] -0.08592201 -0.1566942 -0.10941027
#>  [8,] -0.17455728 -0.2160926 -0.06235645
#>  [9,] -0.16551683 -0.1367584 -0.09728149
#> [10,] -0.10669366 -0.0872110 -0.07221316
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]          [,4]       [,5]          [,6]
#> [1,]  0.12017738 0.04072875 -0.05318146 -1.447986e-02 0.03644839 -0.0006076293
#> [2,] -0.03032964 0.02377744 -0.03879320 -6.757818e-06 0.10583755 -0.0977580316
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.05505044 0.07584562 0.02569380  0.09722416
#> [2,]  0.07533376 0.01014583 0.03733519 -0.09498500
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -3.444145e-05 -3.498675e-05 -0.0001144771 -0.0001267396 -4.953435e-05
#>  [2,]  4.480754e-05  3.022595e-04  0.0003192163  0.0003955709  4.755153e-04
#>  [3,]  3.586398e-05  1.830972e-04  0.0002056495  0.0002621245  3.082809e-04
#>  [4,]  1.677515e-04  9.287813e-04  0.0011108320  0.0012377139  1.146022e-03
#>  [5,]  2.466153e-04  1.295465e-03  0.0016472416  0.0018934912  1.671767e-03
#>  [6,] -3.941922e-05 -2.570439e-05 -0.0001294658 -0.0001269864  9.988423e-06
#>  [7,]  2.922141e-04  1.464347e-03  0.0019591624  0.0022814398  1.892344e-03
#>  [8,] -3.751212e-05 -7.079414e-05 -0.0002016594 -0.0002075309  3.710521e-06
#>  [9,]  1.756826e-04  1.038898e-03  0.0012344667  0.0013745404  1.257850e-03
#> [10,] -2.145583e-04 -8.239556e-04 -0.0012345445 -0.0014009342 -1.016048e-03
#>                 V14           V15          V16          V17          V18
#>  [1,]  3.466563e-05  0.0001057800 0.0002056071 0.0002920942 0.0003369652
#>  [2,]  6.593207e-04  0.0008630200 0.0011104688 0.0012768922 0.0013768409
#>  [3,]  4.390614e-04  0.0005734882 0.0007300993 0.0008061801 0.0008720859
#>  [4,]  1.073685e-03  0.0011087281 0.0011981925 0.0012462829 0.0013151667
#>  [5,]  1.471083e-03  0.0014801595 0.0015675011 0.0015927440 0.0017347066
#>  [6,]  2.126699e-04  0.0003864103 0.0006053316 0.0007838691 0.0008656473
#>  [7,]  1.503005e-03  0.0014149851 0.0014100135 0.0013265643 0.0014930622
#>  [8,]  3.376579e-04  0.0006028106 0.0009050750 0.0011355479 0.0012268226
#>  [9,]  1.172249e-03  0.0012022728 0.0013199750 0.0013870688 0.0014991330
#> [10,] -5.077335e-04 -0.0001999629 0.0001217983 0.0004006180 0.0004710148
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0002382546 -4.189193e-05  9.659868e-05  0.0002087423  0.0002596302
#>  [2,]  0.0013686284  6.426562e-05  1.361862e-03  0.0015312873  0.0016121580
#>  [3,]  0.0009031255  5.201818e-05  9.450426e-04  0.0010590217  0.0011078262
#>  [4,]  0.0016226216  2.185060e-04  2.010455e-03  0.0020015282  0.0019426359
#>  [5,]  0.0023286644  3.292305e-04  2.956853e-03  0.0029438505  0.0029096995
#>  [6,]  0.0006843608 -4.448544e-05  4.466788e-04  0.0006290936  0.0007150292
#>  [7,]  0.0023288986  3.959770e-04  3.087531e-03  0.0029944133  0.0029009766
#>  [8,]  0.0009852369 -4.648007e-05  7.359926e-04  0.0009838336  0.0011141028
#>  [9,]  0.0018708529  2.331553e-04  2.293039e-03  0.0023389625  0.0023282919
#> [10,] -0.0001699018 -2.784583e-04 -9.027684e-04 -0.0005612242 -0.0003596572
#>                 V23          V24          V25          V26          V27
#>  [1,]  0.0003335185 0.0004598982 0.0005577148 0.0005440733 0.0004855439
#>  [2,]  0.0018137036 0.0020563990 0.0021636066 0.0021023251 0.0020966278
#>  [3,]  0.0012501726 0.0014231201 0.0014946270 0.0014800729 0.0014901845
#>  [4,]  0.0020299585 0.0020502460 0.0019835765 0.0019607417 0.0020774540
#>  [5,]  0.0030020012 0.0029557722 0.0027885554 0.0027174227 0.0028885974
#>  [6,]  0.0008799678 0.0011442033 0.0013077287 0.0012512549 0.0011824416
#>  [7,]  0.0028795849 0.0027010182 0.0023971603 0.0023109491 0.0025242802
#>  [8,]  0.0013458594 0.0016834694 0.0018938030 0.0018484896 0.0017604523
#>  [9,]  0.0024362283 0.0024814588 0.0024233007 0.0024035765 0.0025138140
#> [10,] -0.0001267053 0.0003511784 0.0007616454 0.0007569330 0.0005217001
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0005079034 0.0005411994 -3.620835e-05 0.0004448593 0.0003901812
#>  [2,] 0.0021466695 0.0020446927  1.035787e-04 0.0017834520 0.0016103646
#>  [3,] 0.0015063619 0.0014204839  7.512213e-05 0.0012336781 0.0011000251
#>  [4,] 0.0022913387 0.0021381910  2.616504e-04 0.0019373222 0.0017230133
#>  [5,] 0.0031215134 0.0028550701  3.803244e-04 0.0025550515 0.0022690605
#>  [6,] 0.0012165063 0.0012331920 -2.155775e-05 0.0010401770 0.0009391173
#>  [7,] 0.0027969960 0.0025389830  4.358710e-04 0.0022461577 0.0019413912
#>  [8,] 0.0017365922 0.0017163426 -1.360199e-05 0.0014799229 0.0013467958
#>  [9,] 0.0027178000 0.0025503696  2.750576e-04 0.0022921866 0.0020314745
#> [10,] 0.0003523078 0.0004923810 -2.722382e-04 0.0003413802 0.0003409874
#>                 V32          V33          V34          V35          V36
#>  [1,]  0.0002275339 3.139665e-04 0.0003619545 0.0004346271 0.0005437001
#>  [2,]  0.0013452437 1.392439e-03 0.0015319741 0.0014381911 0.0013104330
#>  [3,]  0.0009476562 9.781028e-04 0.0010909678 0.0009889786 0.0008515152
#>  [4,]  0.0016873691 1.589043e-03 0.0015694562 0.0012191437 0.0007760230
#>  [5,]  0.0023227347 2.165421e-03 0.0021779789 0.0015573543 0.0007775431
#>  [6,]  0.0006295700 7.531463e-04 0.0008757286 0.0009710724 0.0011183719
#>  [7,]  0.0021730578 1.998593e-03 0.0020207265 0.0011887414 0.0001429981
#>  [8,]  0.0009483904 1.085323e-03 0.0012515273 0.0013828848 0.0015378210
#>  [9,]  0.0019653334 1.883853e-03 0.0018747362 0.0014532192 0.0009404740
#> [10,] -0.0002133406 7.333086e-05 0.0002416156 0.0007362391 0.0014256700
#>                 V37          V38           V39            V4          V40
#>  [1,]  5.372065e-04 0.0002911476  0.0001972525 -5.313159e-05 2.743294e-04
#>  [2,]  1.195789e-03 0.0009774198  0.0008809211  1.262413e-04 9.532532e-04
#>  [3,]  7.554632e-04 0.0006553273  0.0005913700  9.477814e-05 6.215489e-04
#>  [4,]  6.911590e-04 0.0010445912  0.0012164357  3.282300e-04 1.145369e-03
#>  [5,]  6.112249e-04 0.0013252967  0.0016869029  4.793309e-04 1.584530e-03
#>  [6,]  1.083662e-03 0.0006274846  0.0004445028 -3.535171e-05 5.846041e-04
#>  [7,] -8.343060e-06 0.0011069333  0.0016739247  5.571989e-04 1.544263e-03
#>  [8,]  1.454610e-03 0.0008606048  0.0005749920 -2.342397e-05 6.989488e-04
#>  [9,]  8.359840e-04 0.0012096982  0.0013948283  3.437964e-04 1.336326e-03
#> [10,]  1.462986e-03 0.0002627742 -0.0002887283 -3.554057e-04 1.461161e-05
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0001963299  8.420715e-05  1.989980e-06 -3.724026e-05 -0.0001506293
#>  [2,]  0.0008008370  5.506308e-04  4.109503e-04  3.469580e-04  0.0002447419
#>  [3,]  0.0005273409  3.584462e-04  2.872077e-04  2.516574e-04  0.0001883404
#>  [4,]  0.0011286700  1.039651e-03  9.503963e-04  9.103507e-04  0.0011115731
#>  [5,]  0.0016107743  1.528596e-03  1.446869e-03  1.403193e-03  0.0016566423
#>  [6,]  0.0004196576  1.823304e-04  3.043169e-05 -3.142541e-05 -0.0002276579
#>  [7,]  0.0016410075  1.628714e-03  1.633320e-03  1.621200e-03  0.0019500152
#>  [8,]  0.0004719992  1.682024e-04  4.317588e-06 -7.278453e-05 -0.0003166360
#>  [9,]  0.0013146947  1.194706e-03  1.087537e-03  1.048439e-03  0.0012224589
#> [10,] -0.0002814733 -6.410382e-04 -8.179434e-04 -8.679720e-04 -0.0013712524
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0001212417 -4.117318e-05 -1.339535e-05 -1.016415e-05 -3.923094e-05
#>  [2,]  0.0001819257  2.022840e-04  1.493385e-04  9.184475e-05  1.438493e-04
#>  [3,]  0.0001408145  1.374030e-04  9.643555e-05  5.912707e-05  1.046939e-04
#>  [4,]  0.0008601485  6.218068e-04  4.373502e-04  2.776587e-04  3.326308e-04
#>  [5,]  0.0012542300  9.026762e-04  6.160300e-04  4.023808e-04  5.003052e-04
#>  [6,] -0.0001810690 -4.194695e-05 -6.918869e-06 -9.578759e-06 -1.435856e-05
#>  [7,]  0.0014667151  1.032840e-03  6.926890e-04  4.619516e-04  5.862907e-04
#>  [8,] -0.0002480020 -8.071114e-05 -3.405071e-05 -3.076500e-05 -4.703552e-06
#>  [9,]  0.0009312632  6.855122e-04  4.847843e-04  3.097869e-04  3.560071e-04
#> [10,] -0.0010731724 -6.203674e-04 -3.837487e-04 -2.606689e-04 -3.350571e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.299861e-06 -3.953225e-06 -9.600195e-06  2.198338e-07  8.375271e-07
#>  [2,]  4.460651e-05  2.355392e-05  2.020233e-05  2.930197e-05  2.855243e-05
#>  [3,]  2.898880e-05  1.738225e-05  1.601033e-05  2.196028e-05  2.131216e-05
#>  [4,]  1.009422e-04  6.326298e-05  6.176104e-05  5.010158e-05  4.424170e-05
#>  [5,]  1.377850e-04  9.471515e-05  9.210172e-05  7.411415e-05  6.385458e-05
#>  [6,]  7.206858e-06 -3.231736e-06 -9.762275e-06  6.499453e-06  7.812618e-06
#>  [7,]  1.441980e-04  1.113789e-04  1.081551e-04  8.181157e-05  6.674255e-05
#>  [8,]  9.030680e-06 -5.287768e-06 -9.636788e-06  1.073939e-05  1.352603e-05
#>  [9,]  1.108512e-04  7.176527e-05  6.705465e-05  5.510629e-05  4.810752e-05
#> [10,] -7.073308e-05 -6.425823e-05 -7.181280e-05 -3.276670e-05 -2.518018e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  3.632186e-07  4.172814e-06  4.465323e-06 -1.666796e-06 -1.019994e-06
#>  [2,]  2.162181e-05  1.963712e-05  2.183617e-05  1.107687e-05  1.638984e-05
#>  [3,]  1.573606e-05  1.312934e-05  1.377197e-05  7.301724e-06  1.189019e-05
#>  [4,]  3.543639e-05  2.777670e-05  2.815851e-05  2.931534e-05  3.085904e-05
#>  [5,]  4.703177e-05  3.826635e-05  3.436321e-05  3.811497e-05  4.399912e-05
#>  [6,]  6.092367e-06  9.363963e-06  1.210803e-05 -3.214202e-07  2.059139e-06
#>  [7,]  4.619372e-05  3.864262e-05  2.987056e-05  4.130408e-05  4.760336e-05
#>  [8,]  1.084276e-05  1.266453e-05  1.680869e-05  9.830552e-07  5.278257e-06
#>  [9,]  3.744466e-05  3.173612e-05  3.061659e-05  3.036598e-05  3.366521e-05
#> [10,] -1.839885e-05 -7.274988e-06 -1.261280e-06 -2.530691e-05 -2.331922e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.217701e-05  1.618077e-07  6.210997e-05  0.0000299302 -4.209028e-05
#>  [2,]  2.844787e-04  1.217924e-05  3.304277e-04  0.0003138919  2.903068e-04
#>  [3,]  1.879203e-04  8.455024e-06  2.161100e-04  0.0002053017  1.922432e-04
#>  [4,]  4.367153e-04  2.132391e-05  4.577091e-04  0.0005572871  7.826271e-04
#>  [5,]  6.283485e-04  2.811102e-05  6.296846e-04  0.0008021445  1.098972e-03
#>  [6,]  1.120599e-04  3.116526e-06  1.633716e-04  0.0001011304 -1.491538e-05
#>  [7,]  6.670677e-04  2.877615e-05  6.431451e-04  0.0008822579  1.217304e-03
#>  [8,]  1.432533e-04  6.102880e-06  2.116985e-04  0.0001204304 -2.043514e-05
#>  [9,]  4.920181e-04  2.286041e-05  5.198322e-04  0.0006359033  8.736624e-04
#> [10,] -1.934606e-04 -1.220446e-05 -9.914785e-05 -0.0003149492 -6.691186e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,] -0.04773895 -0.01728175 -0.004668949 -0.03832485 -0.01967258 -0.03248414
#> [2,]  0.04773895  0.01728175  0.004668949  0.03832485  0.01967258  0.03248414
#>              [,7]        [,8]        [,9]       [,10]
#> [1,] -0.007835735 -0.02016348 -0.02919748 -0.03159814
#> [2,]  0.007835735  0.02016348  0.02919748  0.03159814
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.86041153  0.46831392 -0.02581272  0.69036198  0.49242781  0.52499229
#>  [7]  0.32227923  0.39166125  0.18338566  0.56858403
#> 
#> $B[[2]]
#> [1]  0.13497493 -0.07995827
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0006509081 0.0030037677 0.0020843596 0.0033596721 0.0046310907
#>  [6] 0.0016205853 0.0043052564 0.0023332626 0.0039048363 0.0001548503
#> 
#> $vB[[2]]
#> [1] -0.05893872  0.05893872
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [2,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#>  [3,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#>  [4,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#>  [5,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#>  [6,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [7,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#>  [8,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#>  [9,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [10,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [11,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [12,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [13,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [14,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [15,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [16,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [17,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [18,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [19,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [20,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [21,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [22,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [23,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [24,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [25,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [26,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [27,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [28,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [29,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [30,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [31,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [32,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [33,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [34,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [35,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [36,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [37,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [38,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [39,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [40,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [2,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#>  [3,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#>  [4,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#>  [5,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#>  [6,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [7,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#>  [8,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#>  [9,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [10,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [11,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [12,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [13,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [14,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [15,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [16,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [17,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [18,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [19,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [20,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [21,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [22,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [23,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [24,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [25,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [26,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [27,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [28,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [29,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [30,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [31,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [32,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [33,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [34,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [35,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [36,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [37,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [38,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [39,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [40,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [2,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#>  [3,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#>  [4,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#>  [5,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#>  [6,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [7,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#>  [8,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#>  [9,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [10,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [11,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [12,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [13,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [14,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [15,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [16,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [17,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [18,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [19,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [20,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [21,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [22,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [23,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [24,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [25,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [26,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [27,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [28,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [29,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [30,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [31,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [32,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [33,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [34,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [35,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [36,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [37,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [38,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [39,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [40,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [2,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#>  [3,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#>  [4,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#>  [5,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#>  [6,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [7,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#>  [8,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#>  [9,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [10,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [11,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [12,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [13,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [14,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [15,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [16,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [17,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [18,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [19,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [20,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [21,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [22,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [23,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [24,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [25,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [26,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [27,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [28,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [29,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [30,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [31,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [32,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [33,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [34,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [35,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [36,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [37,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [38,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [39,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [40,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [2,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#>  [3,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#>  [4,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#>  [5,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#>  [6,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [7,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#>  [8,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#>  [9,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [10,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [11,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [12,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [13,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [14,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [15,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [16,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [17,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [18,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [19,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [20,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [21,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [22,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [23,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [24,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [25,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [26,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [27,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [28,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [29,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [30,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [31,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [32,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [33,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [34,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [35,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [36,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [37,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [38,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [39,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [40,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [2,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#>  [3,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#>  [4,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#>  [5,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#>  [6,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [7,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#>  [8,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#>  [9,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [10,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [11,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [12,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [13,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [14,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [15,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [16,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [17,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [18,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [19,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [20,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [21,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [22,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [23,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [24,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [25,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [26,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [27,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [28,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [29,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [30,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [31,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [32,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [33,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [34,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [35,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [36,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [37,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [38,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [39,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [40,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9190882 0.5560443 0.2992520 0.8354246 0.6013334 0.7860355 0.4639688
#>  [2,] 0.9081373 0.5357625 0.2867198 0.7917208 0.5485393 0.7463884 0.3174584
#>  [3,] 0.9314367 0.6151533 0.3781880 0.8599462 0.6613556 0.8208162 0.4785627
#>  [4,] 0.8887860 0.4870253 0.2689728 0.7964623 0.5007384 0.7133480 0.4558831
#>  [5,] 0.9513176 0.7090390 0.4995303 0.8916523 0.7627760 0.8793039 0.5703837
#>  [6,] 0.9008837 0.6595456 0.4555022 0.8350156 0.6973638 0.8075415 0.5454934
#>  [7,] 0.9432073 0.5812118 0.3494283 0.8560393 0.6187804 0.8278373 0.5264891
#>  [8,] 0.9486448 0.6792247 0.4410323 0.8761051 0.7292532 0.8749281 0.5068183
#>  [9,] 0.9269997 0.6119392 0.3744204 0.8294766 0.6534269 0.8148514 0.4436507
#> [10,] 0.8975321 0.5122978 0.2554347 0.7967541 0.5159048 0.7169990 0.3268696
#> [11,] 0.9415750 0.6659545 0.4297940 0.8700169 0.7066226 0.8575919 0.5121813
#> [12,] 0.9261547 0.5655970 0.3017367 0.8073136 0.5901201 0.8152568 0.3403362
#> [13,] 0.9428274 0.6856822 0.4538622 0.8811901 0.7290905 0.8634892 0.5518641
#> [14,] 0.9294924 0.6380517 0.4023742 0.8378453 0.6712542 0.8322633 0.4704999
#> [15,] 0.9095981 0.5172760 0.2932315 0.8421871 0.5433692 0.7379494 0.4724042
#> [16,] 0.9226361 0.6168227 0.4007370 0.8307457 0.6553260 0.8261629 0.4495562
#> [17,] 0.9085012 0.6022074 0.3199715 0.8202387 0.5863683 0.7848846 0.4354475
#> [18,] 0.9224358 0.5864129 0.3539484 0.8301278 0.5988617 0.7955409 0.4992231
#> [19,] 0.9499366 0.6740317 0.4251642 0.8811157 0.7190124 0.8716465 0.5111400
#> [20,] 0.9349656 0.6181060 0.3903980 0.8721632 0.6291681 0.8090789 0.4754295
#> [21,] 0.8717081 0.5124417 0.3030290 0.7938804 0.4959066 0.6871864 0.5206408
#> [22,] 0.8414312 0.5276385 0.3260834 0.7315425 0.5403773 0.7024717 0.4371600
#> [23,] 0.9061997 0.5049829 0.2459399 0.7588019 0.5176916 0.7580807 0.2732319
#> [24,] 0.9196134 0.5595754 0.3241167 0.8412162 0.5658372 0.7637747 0.4831747
#> [25,] 0.9590038 0.5978128 0.2835177 0.8723633 0.5768138 0.8491387 0.4223862
#> [26,] 0.9029071 0.6165368 0.3826160 0.8312615 0.6186478 0.7596494 0.4962945
#> [27,] 0.8866470 0.4891281 0.2646535 0.7928333 0.5053035 0.7312941 0.4542894
#> [28,] 0.9552411 0.6934003 0.4591193 0.8975613 0.7284730 0.8759330 0.5482179
#> [29,] 0.9248631 0.6080544 0.3714327 0.8674864 0.6309968 0.7895211 0.5338714
#> [30,] 0.8990820 0.5071494 0.2705565 0.8065388 0.5578224 0.7482214 0.4447887
#> [31,] 0.8830633 0.5113215 0.3047327 0.7469748 0.5331878 0.7323041 0.3582996
#> [32,] 0.9128101 0.5873424 0.3495072 0.8383096 0.6286579 0.7918644 0.5188310
#> [33,] 0.9101626 0.5346198 0.2869001 0.8193652 0.5618135 0.7609829 0.4864877
#> [34,] 0.9205410 0.5547469 0.3145610 0.8463577 0.5490827 0.7518764 0.4829380
#> [35,] 0.9354533 0.5810972 0.2970880 0.8553549 0.6060969 0.8170818 0.3784443
#> [36,] 0.8065965 0.4383439 0.2383409 0.6477645 0.4542063 0.6109705 0.3516193
#> [37,] 0.9605576 0.6942292 0.4529484 0.9111634 0.7378286 0.8880400 0.6490955
#> [38,] 0.8855314 0.4959209 0.2620616 0.7364998 0.5286222 0.7206423 0.2969258
#> [39,] 0.8663606 0.4462832 0.2181383 0.7321176 0.4462937 0.6587036 0.2603262
#> [40,] 0.9487736 0.6732180 0.4377530 0.8809874 0.7154996 0.8668721 0.5128815
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6071250 0.6959203 0.7678592
#>  [2,] 0.5341562 0.5948138 0.7343583
#>  [3,] 0.6343968 0.7079970 0.7853962
#>  [4,] 0.5642666 0.6969210 0.7005643
#>  [5,] 0.6930028 0.7532656 0.8226801
#>  [6,] 0.6608212 0.6833722 0.7776135
#>  [7,] 0.6260644 0.7517041 0.7769278
#>  [8,] 0.6628460 0.7270383 0.8234491
#>  [9,] 0.6194187 0.6744175 0.7834715
#> [10,] 0.5335563 0.6180255 0.7133266
#> [11,] 0.6667186 0.7238905 0.8132802
#> [12,] 0.5784721 0.6018573 0.8049962
#> [13,] 0.6825649 0.7509288 0.8179536
#> [14,] 0.6344147 0.6740833 0.7928302
#> [15,] 0.5771753 0.7383636 0.7018112
#> [16,] 0.6244943 0.6524058 0.7849182
#> [17,] 0.6008319 0.6669801 0.7674346
#> [18,] 0.6212817 0.6914286 0.7644346
#> [19,] 0.6748813 0.7362918 0.8279922
#> [20,] 0.6367173 0.7204816 0.7851123
#> [21,] 0.5710141 0.7094003 0.6881967
#> [22,] 0.5545431 0.5942606 0.6976947
#> [23,] 0.5445468 0.5513181 0.7735414
#> [24,] 0.6121012 0.7269987 0.7421894
#> [25,] 0.6388949 0.7292570 0.8455235
#> [26,] 0.6036971 0.6992615 0.6993483
#> [27,] 0.5474187 0.7071533 0.6900470
#> [28,] 0.6826327 0.7660334 0.8197195
#> [29,] 0.6340010 0.7434344 0.7726444
#> [30,] 0.5612216 0.6869579 0.7324361
#> [31,] 0.5302721 0.5766523 0.6986963
#> [32,] 0.6259721 0.7205699 0.7685776
#> [33,] 0.5751800 0.7359341 0.7216667
#> [34,] 0.6093121 0.7405969 0.7417771
#> [35,] 0.6037505 0.7086714 0.7873759
#> [36,] 0.4890104 0.5434896 0.6220834
#> [37,] 0.7006526 0.8193227 0.8249281
#> [38,] 0.5291547 0.5571160 0.7195765
#> [39,] 0.5306149 0.5496181 0.7107030
#> [40,] 0.6653179 0.7423518 0.8072134
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5031217 0.4968783
#>  [2,] 0.5108213 0.4891787
#>  [3,] 0.5006470 0.4993530
#>  [4,] 0.5040609 0.4959391
#>  [5,] 0.4942079 0.5057921
#>  [6,] 0.4969588 0.5030412
#>  [7,] 0.4995906 0.5004094
#>  [8,] 0.4984931 0.5015069
#>  [9,] 0.5031897 0.4968103
#> [10,] 0.5100685 0.4899315
#> [11,] 0.4986912 0.5013088
#> [12,] 0.5115177 0.4884823
#> [13,] 0.4959899 0.5040101
#> [14,] 0.5018305 0.4981695
#> [15,] 0.5008609 0.4991391
#> [16,] 0.5032112 0.4967888
#> [17,] 0.5048878 0.4951122
#> [18,] 0.5017447 0.4982553
#> [19,] 0.4986687 0.5013313
#> [20,] 0.5010137 0.4989863
#> [21,] 0.5005527 0.4994473
#> [22,] 0.5055325 0.4944675
#> [23,] 0.5165564 0.4834436
#> [24,] 0.5015780 0.4984220
#> [25,] 0.5075535 0.4924465
#> [26,] 0.4980799 0.5019201
#> [27,] 0.5035838 0.4964162
#> [28,] 0.4957655 0.5042345
#> [29,] 0.4979758 0.5020242
#> [30,] 0.5041920 0.4958080
#> [31,] 0.5094529 0.4905471
#> [32,] 0.4995457 0.5004543
#> [33,] 0.5011897 0.4988103
#> [34,] 0.5016508 0.4983492
#> [35,] 0.5064229 0.4935771
#> [36,] 0.5106383 0.4893617
#> [37,] 0.4910715 0.5089285
#> [38,] 0.5133241 0.4866759
#> [39,] 0.5165668 0.4834332
#> [40,] 0.4977411 0.5022589
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]        [,2]         [,3]      [,4]         [,5]      [,6]
#>  [1,] 2.430022  0.22512331 -0.850862071 1.6245713  0.411023934 1.3011917
#>  [2,] 2.291100  0.14329475 -0.911368871 1.3353291  0.194770682 1.0794423
#>  [3,] 2.608970  0.46902646 -0.497246527 1.8148432  0.669341061 1.5218871
#>  [4,] 2.078401 -0.05191027 -0.999840470 1.3643287  0.002953608 0.9117002
#>  [5,] 2.972530  0.89072121 -0.001878735 2.1077311  1.167959395 1.9858544
#>  [6,] 2.207082  0.66126989 -0.178463285 1.6215991  0.834775679 1.4341138
#>  [7,] 2.809879  0.32774974 -0.621553169 1.7827763  0.484375016 1.5703764
#>  [8,] 2.916269  0.75021127 -0.236973746 1.9560524  0.990836883 1.9452528
#>  [9,] 2.541490  0.45547074 -0.513299418 1.5819224  0.634136760 1.4818468
#> [10,] 2.170099  0.04920128 -1.069833581 1.3661296  0.063640749 0.9296242
#> [11,] 2.779810  0.68994432 -0.282691826 1.9011081  0.879036802 1.7954316
#> [12,] 2.529069  0.26390897 -0.839041369 1.4326482  0.364461751 1.4845364
#> [13,] 2.802809  0.78000975 -0.185077679 2.0037489  0.990013044 1.8445776
#> [14,] 2.578918  0.56691789 -0.395582164 1.6422825  0.713863019 1.6017533
#> [15,] 2.308738  0.06913148 -0.879740615 1.6745918  0.173913729 1.0353375
#> [16,] 2.478714  0.47608380 -0.402395009 1.5909212  0.642533821 1.5586731
#> [17,] 2.295471  0.41467108 -0.753902775 1.5179653  0.348972068 1.2943621
#> [18,] 2.475911  0.34915608 -0.601728203 1.5865335  0.400724635 1.3586542
#> [19,] 2.943106  0.72647693 -0.301608876 2.0030378  0.939568214 1.9155962
#> [20,] 2.665594  0.48151683 -0.445639345 1.9202220  0.528649531 1.4440364
#> [21,] 1.916146  0.04977689 -0.832915329 1.3484764 -0.016374145 0.7869981
#> [22,] 1.668915  0.11066691 -0.725953137 1.0024625  0.161861676 0.8590960
#> [23,] 2.268091  0.01993235 -1.120384696 1.1461222  0.070795837 1.1421856
#> [24,] 2.437106  0.23943904 -0.734916792 1.6673051  0.264886893 1.1734866
#> [25,] 3.152416  0.39636022 -0.927078888 1.9220178  0.309707281 1.7278617
#> [26,] 2.229951  0.47487447 -0.478459364 1.5945946  0.483812858 1.1507584
#> [27,] 2.056940 -0.04349462 -1.021920601 1.3420894  0.021214985 1.0011980
#> [28,] 3.060673  0.81606466 -0.163888647 2.1704173  0.986889088 1.9544683
#> [29,] 2.510334  0.43914111 -0.526075442 1.8789145  0.536495543 1.3220409
#> [30,] 2.187066  0.02859944 -0.991800717 1.4276751  0.232328841 1.0891486
#> [31,] 2.021764  0.04529360 -0.824861508 1.0825423  0.132946578 1.0063443
#> [32,] 2.348439  0.35298990 -0.621206171 1.6457037  0.526463425 1.3362004
#> [33,] 2.315622  0.13870130 -0.910487541 1.5120523  0.248525469 1.1580760
#> [34,] 2.449720  0.21986923 -0.778881630 1.7063149  0.196964991 1.1086450
#> [35,] 2.673643  0.32727908 -0.861203322 1.7772337  0.430935027 1.4967001
#> [36,] 1.428045 -0.24788586 -1.161796947 0.6092273 -0.183689749 0.4513934
#> [37,] 3.192673  0.81996632 -0.188764842 2.3279233  1.034712860 2.0708751
#> [38,] 2.045887 -0.01631682 -1.035280774 1.0278547  0.114614123 0.9476500
#> [39,] 1.869156 -0.21569949 -1.276548639 1.0053932 -0.215657044 0.6575225
#> [40,] 2.918916  0.72277585 -0.250286616 2.0018144  0.922246203 1.8735812
#>               [,7]        [,8]      [,9]     [,10]
#>  [1,] -0.144375057  0.43524326 0.8279454 1.1962621
#>  [2,] -0.765476598  0.13683773 0.3839017 1.0168487
#>  [3,] -0.085801751  0.55112616 0.8856759 1.2973945
#>  [4,] -0.176927533  0.25849640 0.8326788 0.8499865
#>  [5,]  0.283416731  0.81419535 1.1161055 1.5346117
#>  [6,]  0.182478101  0.66695588 0.7693128 1.2518125
#>  [7,]  0.106055550  0.51536946 1.1077216 1.2478517
#>  [8,]  0.027274711  0.67600280 0.9796477 1.5398928
#>  [9,] -0.226358958  0.48708171 0.7282333 1.2860125
#> [10,] -0.722377565  0.13442723 0.4811757 0.9115958
#> [11,]  0.048734734  0.69338093 0.9638424 1.4714664
#> [12,] -0.661796315  0.31650436 0.4132097 1.4178186
#> [13,]  0.208205224  0.76558432 1.1035721 1.5025441
#> [14,] -0.118137436  0.55120301 0.7267117 1.3420706
#> [15,] -0.110495386  0.31118836 1.0374807 0.8559377
#> [16,] -0.202463886  0.50866849 0.6296309 1.2945609
#> [17,] -0.259659305  0.40893273 0.6945582 1.1938816
#> [18,] -0.003107566  0.49499208 0.8068064 1.1771480
#> [19,]  0.044567355  0.73034669 1.0267833 1.5714641
#> [20,] -0.098361426  0.56114462 0.9468519 1.2957111
#> [21,]  0.082610148  0.28598994 0.8924733 0.7917021
#> [22,] -0.252696170  0.21904388 0.3816071 0.8363441
#> [23,] -0.978286492  0.17866108 0.2059979 1.2284169
#> [24,] -0.067326500  0.45615307 0.9794480 1.0573792
#> [25,] -0.312985284  0.57057105 0.9908561 1.6999137
#> [26,] -0.014822392  0.42089361 0.8437837 0.8441962
#> [27,] -0.183354223  0.19024666 0.8815983 0.8003391
#> [28,]  0.193472731  0.76589739 1.1860472 1.5144481
#> [29,]  0.135693370  0.54941989 1.0638961 1.2233035
#> [30,] -0.221749462  0.24612141 0.7859351 1.0070179
#> [31,] -0.582752085  0.12123669 0.3090456 0.8410976
#> [32,]  0.075359752  0.51497559 0.9472904 1.2002972
#> [33,] -0.054062296  0.30301759 1.0249419 0.9527443
#> [34,] -0.068274355  0.44442152 1.0490732 1.0552258
#> [35,] -0.496156556  0.42111704 0.8889399 1.3091797
#> [36,] -0.611928916 -0.04396553 0.1743991 0.4984008
#> [37,]  0.615065778  0.85040761 1.5117654 1.5500995
#> [38,] -0.861980164  0.11675132 0.2294655 0.9423619
#> [39,] -1.044273873  0.12261313 0.1991277 0.8988007
#> [40,]  0.051537430  0.68708363 1.0582283 1.4320041
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1308835 0.11839667
#>  [2,] 0.1363799 0.09308780
#>  [3,] 0.1288679 0.12627997
#>  [4,] 0.1252993 0.10905518
#>  [5,] 0.1233586 0.14652826
#>  [6,] 0.1224206 0.13458553
#>  [7,] 0.1256615 0.12729926
#>  [8,] 0.1287141 0.13474176
#>  [9,] 0.1317153 0.11895626
#> [10,] 0.1349439 0.09466459
#> [11,] 0.1282785 0.13351381
#> [12,] 0.1407510 0.09467202
#> [13,] 0.1256300 0.14167085
#> [14,] 0.1299338 0.12261165
#> [15,] 0.1232158 0.11977215
#> [16,] 0.1296651 0.11682022
#> [17,] 0.1321060 0.11255430
#> [18,] 0.1269598 0.11998114
#> [19,] 0.1299402 0.13526529
#> [20,] 0.1279536 0.12389876
#> [21,] 0.1190569 0.11684617
#> [22,] 0.1247777 0.10264657
#> [23,] 0.1447945 0.07854464
#> [24,] 0.1266771 0.12036518
#> [25,] 0.1405784 0.11036205
#> [26,] 0.1209590 0.12863955
#> [27,] 0.1235539 0.10921841
#> [28,] 0.1254155 0.14235384
#> [29,] 0.1242831 0.13237995
#> [30,] 0.1283515 0.11158291
#> [31,] 0.1304585 0.09264224
#> [32,] 0.1259286 0.12774567
#> [33,] 0.1250967 0.12033775
#> [34,] 0.1268099 0.12020658
#> [35,] 0.1372082 0.11151508
#> [36,] 0.1281889 0.08562920
#> [37,] 0.1200298 0.15574759
#> [38,] 0.1388432 0.08553417
#> [39,] 0.1415584 0.07526709
#> [40,] 0.1271875 0.13622316
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5031217  0.5031217
#>  [2,]  0.4891787 -0.4891787
#>  [3,]  0.4993530 -0.4993530
#>  [4,] -0.5040609  0.5040609
#>  [5,] -0.4942079  0.4942079
#>  [6,] -0.4969588  0.4969588
#>  [7,] -0.4995906  0.4995906
#>  [8,]  0.5015069 -0.5015069
#>  [9,]  0.4968103 -0.4968103
#> [10,]  0.4899315 -0.4899315
#> [11,] -0.4986912  0.4986912
#> [12,]  0.4884823 -0.4884823
#> [13,] -0.4959899  0.4959899
#> [14,]  0.4981695 -0.4981695
#> [15,]  0.4991391 -0.4991391
#> [16,]  0.4967888 -0.4967888
#> [17,] -0.5048878  0.5048878
#> [18,] -0.5017447  0.5017447
#> [19,] -0.4986687  0.4986687
#> [20,] -0.5010137  0.5010137
#> [21,]  0.4994473 -0.4994473
#> [22,]  0.4944675 -0.4944675
#> [23,]  0.4834436 -0.4834436
#> [24,]  0.4984220 -0.4984220
#> [25,] -0.5075535  0.5075535
#> [26,]  0.5019201 -0.5019201
#> [27,]  0.4964162 -0.4964162
#> [28,]  0.5042345 -0.5042345
#> [29,]  0.5020242 -0.5020242
#> [30,] -0.5041920  0.5041920
#> [31,]  0.4905471 -0.4905471
#> [32,]  0.5004543 -0.5004543
#> [33,]  0.4988103 -0.4988103
#> [34,]  0.4983492 -0.4983492
#> [35,]  0.4935771 -0.4935771
#> [36,]  0.4893617 -0.4893617
#> [37,] -0.4910715  0.4910715
#> [38,]  0.4866759 -0.4866759
#> [39,] -0.5165668  0.5165668
#> [40,]  0.5022589 -0.5022589
#> 
#> $L
#> [1] 0.6897894 0.6786194 0.7826596 0.7622798 0.7730138 0.6891361
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174