Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12           V13          V14
#>  [1,] -0.2785297 -0.14813497 -0.10762295 -0.16223233  0.0324824030 -0.048871787
#>  [2,] -0.3938757 -0.05996642 -0.18233359 -0.03743024 -0.0639098992 -0.090538049
#>  [3,] -0.3970097 -0.05648521 -0.12028912  0.01431189 -0.0001142739 -0.115307062
#>  [4,] -0.2339025 -0.16017844 -0.16835322 -0.22432758 -0.1569695073 -0.017464228
#>  [5,] -0.2354143 -0.20411188 -0.13728543 -0.07969332 -0.1391602040 -0.003382974
#>  [6,] -0.3168223 -0.18973788 -0.15771220 -0.03188452 -0.0526873977  0.011284819
#>  [7,] -0.3324067 -0.07081867  0.01856763 -0.12816313 -0.0039589333 -0.037433099
#>  [8,] -0.2496390 -0.10586057 -0.10864056 -0.05068770 -0.0593155616 -0.065819815
#>  [9,] -0.2704099 -0.05635024 -0.11383014  0.01230511 -0.0712955166 -0.022546695
#> [10,] -0.2516307 -0.20057453 -0.03262002 -0.14496508 -0.1678399341 -0.012567374
#>                V15          V16         V17         V18         V19         V2
#>  [1,] -0.137048045 -0.071004840 -0.04131912 -0.03914824  0.06321700 -0.2156392
#>  [2,] -0.118292727 -0.089646513 -0.05285438  0.03784706 -0.13652057 -0.3711771
#>  [3,]  0.042474579  0.109661114  0.01102532  0.09988075  0.19007893 -0.3551450
#>  [4,] -0.088149332 -0.043697782  0.10628717  0.05868071  0.13253190 -0.2524954
#>  [5,]  0.002900951 -0.049198679 -0.07902521 -0.02964867  0.06632613 -0.2295771
#>  [6,] -0.124230116 -0.022090619 -0.04570408  0.09342589  0.11987395 -0.3236562
#>  [7,] -0.092638621 -0.005858684  0.05738326  0.01498691  0.04754724 -0.2552443
#>  [8,] -0.100696141 -0.064190193 -0.16360327 -0.14421017 -0.09495547 -0.2042644
#>  [9,]  0.019676424  0.003647288  0.06500389  0.01098102  0.01536132 -0.3224832
#> [10,]  0.010362342 -0.003153817 -0.03744515 -0.05418305  0.02700255 -0.3094975
#>               V20         V21         V22         V23        V24        V25
#>  [1,]  0.07268821  0.06739695  0.11456051  0.20232856 0.24227505 0.22302215
#>  [2,] -0.10608426 -0.04779891 -0.01969512  0.07383083 0.07688268 0.04009326
#>  [3,]  0.21928069  0.20982397  0.27627116  0.32380718 0.21987817 0.32359003
#>  [4,]  0.07364013  0.10690541  0.24506921  0.10723193 0.24479120 0.11352094
#>  [5,] -0.02283312  0.03917405  0.01574716  0.04912638 0.06850058 0.02109529
#>  [6,]  0.13930944  0.01371568  0.12284697  0.16499853 0.20311403 0.10384334
#>  [7,]  0.15480909  0.24367951  0.25572904  0.15056537 0.19038690 0.22680518
#>  [8,] -0.12128664  0.05992550  0.04147935 -0.02763676 0.06142996 0.03915183
#>  [9,]  0.08777790  0.12823407  0.22937065  0.22964321 0.12404629 0.12125866
#> [10,]  0.14035077  0.13215774  0.07845286  0.08843401 0.12072721 0.09085917
#>              V26        V27         V28        V29         V3         V30
#>  [1,] 0.03989487 0.07578898  0.04225499 0.17294187 -0.3003944  0.01326221
#>  [2,] 0.07551677 0.06452261  0.17741925 0.15996416 -0.2875162  0.07818728
#>  [3,] 0.27354330 0.14540306  0.22501093 0.05997118 -0.3899864  0.16735049
#>  [4,] 0.21244950 0.21255158  0.16404475 0.10928191 -0.3152581  0.02400339
#>  [5,] 0.18958443 0.22703310  0.05130978 0.10134720 -0.2485625  0.02740511
#>  [6,] 0.16811758 0.19929247  0.10835553 0.11879132 -0.2224227  0.03325403
#>  [7,] 0.26563386 0.16647310  0.18401761 0.11309055 -0.2568689  0.20610972
#>  [8,] 0.06234165 0.04677008 -0.01075882 0.04336567 -0.1800220 -0.08013653
#>  [9,] 0.17293023 0.17078597  0.17677165 0.13383985 -0.3528256  0.07657450
#> [10,] 0.10156315 0.12430066  0.19684409 0.08935184 -0.2377959  0.08203992
#>               V31          V32         V33         V34          V35
#>  [1,] -0.01442986 -0.019635350  0.01163665  0.01423303 -0.079772276
#>  [2,] -0.05575390  0.043922320 -0.12971330 -0.15457200 -0.120652933
#>  [3,]  0.02379221  0.034036455  0.08071869  0.02924177 -0.031561617
#>  [4,]  0.03072010 -0.136759805  0.02750468 -0.06409689 -0.105492039
#>  [5,]  0.09449195  0.083568968 -0.08434036 -0.05039352 -0.004882389
#>  [6,]  0.11868857  0.104434937  0.03431540  0.14919274  0.082920426
#>  [7,] -0.03533830  0.021419966  0.01571208 -0.02420170 -0.073510807
#>  [8,]  0.02685191 -0.007167609 -0.01545337 -0.03433418 -0.047370263
#>  [9,] -0.12095008 -0.166010788 -0.15758817 -0.04327521 -0.092121900
#> [10,]  0.03442685  0.107007773  0.11684433 -0.05570257  0.037238722
#>                V36         V37         V38         V39         V4          V40
#>  [1,] -0.143662084 -0.08743486 -0.08675974  0.01705245 -0.2494718 -0.057285139
#>  [2,] -0.050271913 -0.01693668 -0.03827784 -0.10885893 -0.2536954 -0.116179492
#>  [3,] -0.020460293 -0.03072349  0.01647549 -0.04425108 -0.3385295 -0.012612291
#>  [4,] -0.083363304 -0.13185967 -0.09471927 -0.16174015 -0.1988701 -0.107440266
#>  [5,]  0.051141774  0.04605734 -0.03748523 -0.05739492 -0.1769357 -0.203601982
#>  [6,] -0.002863593 -0.09293066 -0.09113011 -0.09806750 -0.2172229 -0.050070106
#>  [7,]  0.039209152  0.02273212 -0.06291066 -0.02194180 -0.3726376  0.027999792
#>  [8,] -0.082090993 -0.09513354 -0.05275200 -0.11216203 -0.3343946  0.001771928
#>  [9,] -0.086988984 -0.15453967 -0.13267431 -0.08402996 -0.2036353 -0.100062457
#> [10,] -0.037333944  0.10136311  0.07198616  0.05498259 -0.3306896 -0.052571361
#>                V41          V42         V43         V44         V45         V46
#>  [1,] -0.051268383 -0.003200777  0.04596802 -0.03849030 -0.07694143 -0.05410494
#>  [2,] -0.049986037 -0.020476277 -0.05683862 -0.14773992 -0.07474297 -0.09757313
#>  [3,] -0.008591887 -0.032006472 -0.11097327 -0.10283455 -0.15950765 -0.13043285
#>  [4,] -0.117191546 -0.148618981 -0.07759379 -0.12107498 -0.11877611 -0.28038509
#>  [5,] -0.075922803 -0.063211592 -0.14844356 -0.05708263 -0.13647585 -0.15861730
#>  [6,] -0.052465847 -0.055402919 -0.04493320 -0.15438863 -0.05197069 -0.16775024
#>  [7,] -0.047288340 -0.160989645 -0.10027633 -0.05478674 -0.12134801 -0.08394659
#>  [8,] -0.040800275 -0.144793354 -0.10413710 -0.24678526 -0.16226102 -0.07714349
#>  [9,] -0.076153849 -0.127898629 -0.18433284 -0.04277404 -0.20651053 -0.18666050
#> [10,] -0.020564656 -0.016722147 -0.05255998 -0.02011598 -0.14141258 -0.18858132
#>               V47        V48        V49         V5        V50        V51
#>  [1,] -0.22664839 -0.1290632 -0.2342136 -0.1788872 -0.3164739 -0.3497025
#>  [2,] -0.17794746 -0.2158326 -0.1839723 -0.2285063 -0.3697193 -0.3775321
#>  [3,] -0.15147677 -0.2673975 -0.3497917 -0.2723322 -0.3641844 -0.2921992
#>  [4,] -0.18626314 -0.2230883 -0.2607800 -0.1874911 -0.2466749 -0.2182194
#>  [5,] -0.19378943 -0.2536173 -0.3355159 -0.2773036 -0.3397827 -0.2807233
#>  [6,] -0.14290158 -0.2963887 -0.2396330 -0.2291721 -0.3862801 -0.3527664
#>  [7,] -0.22200860 -0.1581720 -0.2217363 -0.2410355 -0.2958260 -0.4320786
#>  [8,] -0.14156008 -0.2041187 -0.3674142 -0.1773259 -0.2800851 -0.3845160
#>  [9,] -0.27844097 -0.2490769 -0.3143496 -0.2142767 -0.3064362 -0.2630259
#> [10,] -0.09831307 -0.1344148 -0.3219758 -0.2592187 -0.2047401 -0.2331641
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3671123 -0.3889727 -0.2951549 -0.3778710 -0.3916106 -0.4431337
#>  [2,] -0.3824331 -0.2761442 -0.3077449 -0.3728002 -0.2815337 -0.3946516
#>  [3,] -0.4651495 -0.3879143 -0.4976869 -0.3918910 -0.3690966 -0.4797647
#>  [4,] -0.2552180 -0.3664176 -0.3633420 -0.2348276 -0.3096197 -0.2398874
#>  [5,] -0.2513071 -0.3711171 -0.2486742 -0.3349196 -0.2636873 -0.1989109
#>  [6,] -0.3232822 -0.2972169 -0.2678765 -0.2984554 -0.3498707 -0.3438223
#>  [7,] -0.4464074 -0.4189953 -0.3790690 -0.4120583 -0.4162437 -0.4282729
#>  [8,] -0.3668554 -0.2590780 -0.3778405 -0.3252983 -0.4115962 -0.4012878
#>  [9,] -0.3181057 -0.3713937 -0.2828732 -0.3122198 -0.3554716 -0.3193598
#> [10,] -0.2224067 -0.2940404 -0.2435974 -0.2797528 -0.1862258 -0.1921679
#>              V58        V59          V6        V60         V7         V8
#>  [1,] -0.4425447 -0.4070453 -0.19545006 -0.4202401 -0.2584838 -0.2549421
#>  [2,] -0.2434715 -0.3875669 -0.09826677 -0.2748340 -0.1014133 -0.1256623
#>  [3,] -0.3367314 -0.3427211 -0.30127597 -0.4769369 -0.1988953 -0.1527283
#>  [4,] -0.2922816 -0.2690546 -0.22164196 -0.3077823 -0.2423477 -0.1922102
#>  [5,] -0.2288362 -0.2805272 -0.21089129 -0.3288909 -0.2764559 -0.1913969
#>  [6,] -0.3624935 -0.3603039 -0.23514550 -0.2194409 -0.1087297 -0.2203854
#>  [7,] -0.4791120 -0.3358272 -0.25361772 -0.4077666 -0.1405085 -0.2040201
#>  [8,] -0.3000517 -0.3372313 -0.12591851 -0.3629947 -0.3049760 -0.1953630
#>  [9,] -0.4214705 -0.3640192 -0.29637618 -0.3104085 -0.1143854 -0.1977671
#> [10,] -0.2752003 -0.3555946 -0.16785671 -0.3214044 -0.1833422 -0.1201496
#>                V9
#>  [1,] -0.15175342
#>  [2,] -0.24386582
#>  [3,] -0.11166116
#>  [4,] -0.13028994
#>  [5,] -0.08893327
#>  [6,] -0.21033493
#>  [7,] -0.06640406
#>  [8,] -0.20805893
#>  [9,] -0.14635356
#> [10,] -0.08994788
#> 
#> $W[[2]]
#>             [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#> [1,] -0.05696793 -0.11302678 -0.1680826 -0.1897507 -0.06236936 -0.08028397
#> [2,]  0.16597974  0.06604523  0.1565745  0.1655813  0.11756198  0.23764718
#>            [,7]        [,8]        [,9]       [,10]
#> [1,] -0.1911769 -0.21608226 -0.10188796 -0.17759143
#> [2,]  0.1203826  0.02396441  0.09945128  0.07719044
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  9.626629e-05  0.0005365089  4.799189e-04  4.648096e-04  0.0006890123
#>  [2,]  4.967738e-05  0.0002578734  1.925756e-04  1.925469e-04  0.0003722453
#>  [3,]  6.177383e-05  0.0003278625  3.223375e-04  2.982212e-04  0.0003747776
#>  [4,]  4.566147e-05  0.0001946122  3.138780e-04  3.050829e-04  0.0002044075
#>  [5,]  6.090710e-05  0.0003300959  1.976831e-04  1.982989e-04  0.0004849879
#>  [6,]  7.177958e-05  0.0003721377  3.699687e-04  3.454464e-04  0.0004386557
#>  [7,]  8.700306e-05  0.0004413172  4.518891e-04  4.311307e-04  0.0005299211
#>  [8,] -2.649771e-05 -0.0001287961 -9.013994e-05 -9.699882e-05 -0.0002033877
#>  [9,]  1.049571e-04  0.0005699599  5.136660e-04  5.202407e-04  0.0007523744
#> [10,]  7.877505e-05  0.0004189794  3.782772e-04  3.661995e-04  0.0005403102
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0012800377  0.0016907155  0.0022504870  0.0022962943  0.0024067988
#>  [2,]  0.0008424608  0.0011662460  0.0015882168  0.0016357453  0.0017031738
#>  [3,]  0.0005512613  0.0006800656  0.0008501466  0.0008249314  0.0008737363
#>  [4,] -0.0002126085 -0.0004103758 -0.0006563270 -0.0007384803 -0.0007546675
#>  [5,]  0.0012855416  0.0018143928  0.0024991801  0.0025832335  0.0026808551
#>  [6,]  0.0006379166  0.0008007565  0.0010086544  0.0009845767  0.0010305813
#>  [7,]  0.0007311680  0.0008969127  0.0011197799  0.0010815500  0.0011455395
#>  [8,] -0.0005173543 -0.0007105794 -0.0009631685 -0.0009784251 -0.0010184453
#>  [9,]  0.0013852804  0.0017937031  0.0023916437  0.0024403337  0.0026093999
#> [10,]  0.0009793849  0.0012958843  0.0017092993  0.0017183924  0.0017996763
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0021415361  1.717007e-04  2.101869e-03  0.0025697221  0.0029909024
#>  [2,]  0.0014576247  8.758626e-05  1.385822e-03  0.0017101703  0.0019985376
#>  [3,]  0.0008455385  1.019285e-04  9.625626e-04  0.0011390689  0.0012322579
#>  [4,] -0.0003698714  5.397569e-05 -6.758067e-06 -0.0001794695 -0.0004377599
#>  [5,]  0.0022031303  1.212196e-04  1.986276e-03  0.0025206428  0.0030300708
#>  [6,]  0.0010039003  1.192929e-04  1.136880e-03  0.0013437155  0.0014643593
#>  [7,]  0.0011488961  1.396886e-04  1.321966e-03  0.0015400073  0.0016576543
#>  [8,] -0.0008905376 -3.735692e-05 -8.194672e-04 -0.0010037198 -0.0011879713
#>  [9,]  0.0023189360  1.843452e-04  2.286402e-03  0.0027998562  0.0032671365
#> [10,]  0.0016150266  1.385890e-04  1.633603e-03  0.0019947236  0.0022978123
#>                 V23          V24          V25          V26          V27
#>  [1,]  0.0036743452  0.004157148  0.004045053  0.003890788  0.003904598
#>  [2,]  0.0025542906  0.002970254  0.002941924  0.002815833  0.002768122
#>  [3,]  0.0014227447  0.001560433  0.001502928  0.001477291  0.001510362
#>  [4,] -0.0009541178 -0.001390274 -0.001482106 -0.001357803 -0.001174309
#>  [5,]  0.0039767943  0.004688575  0.004652982  0.004417752  0.004313230
#>  [6,]  0.0016648651  0.001790705  0.001702063  0.001655334  0.001710212
#>  [7,]  0.0018755835  0.002014412  0.001915522  0.001876768  0.001943448
#>  [8,] -0.0015961219 -0.001903011 -0.001899650 -0.001845622 -0.001790522
#>  [9,]  0.0040650295  0.004628435  0.004513733  0.004336824  0.004326442
#> [10,]  0.0028042438  0.003159747  0.003066544  0.002936818  0.002948416
#>                 V28          V29            V3          V30          V31
#>  [1,]  0.0037426700  0.003557617  1.856547e-04  0.003243577  0.003017763
#>  [2,]  0.0026240394  0.002577575  1.025209e-04  0.002389730  0.002282564
#>  [3,]  0.0015194404  0.001412992  9.826459e-05  0.001277686  0.001097571
#>  [4,] -0.0009080358 -0.001068763  3.240719e-05 -0.001070166 -0.001286646
#>  [5,]  0.0040266115  0.003973173  1.458996e-04  0.003692843  0.003609992
#>  [6,]  0.0017226446  0.001569899  1.173536e-04  0.001409975  0.001202026
#>  [7,]  0.0019720273  0.001806314  1.346109e-04  0.001617408  0.001369951
#>  [8,] -0.0016419162 -0.001638811 -5.344312e-05 -0.001521172 -0.001500567
#>  [9,]  0.0041574452  0.003999189  1.962663e-04  0.003642868  0.003404219
#> [10,]  0.0028482023  0.002703668  1.460933e-04  0.002454619  0.002246553
#>                 V32           V33           V34          V35          V36
#>  [1,]  0.0024845563  0.0025255730  0.0024270095  0.002820078  0.003052048
#>  [2,]  0.0018750567  0.0019660245  0.0019167741  0.002176496  0.002296215
#>  [3,]  0.0009378073  0.0009226768  0.0009009858  0.001036525  0.001107005
#>  [4,] -0.0009810139 -0.0011407653 -0.0010655972 -0.001235910 -0.001364846
#>  [5,]  0.0029397710  0.0031014868  0.0030012490  0.003448201  0.003685775
#>  [6,]  0.0010201078  0.0010035259  0.0009862794  0.001157717  0.001247305
#>  [7,]  0.0011766732  0.0011370813  0.0011067958  0.001273062  0.001349908
#>  [8,] -0.0012199146 -0.0012549025 -0.0011903564 -0.001300327 -0.001330882
#>  [9,]  0.0028283487  0.0028784256  0.0027550502  0.003172701  0.003421964
#> [10,]  0.0018514211  0.0018869928  0.0018247999  0.002122723  0.002288788
#>                V37           V38           V39            V4           V40
#>  [1,]  0.002838206  0.0019251186  0.0012405718  1.838654e-04  0.0014149515
#>  [2,]  0.002094242  0.0013611275  0.0008439956  9.421123e-05  0.0010135946
#>  [3,]  0.001024682  0.0007680627  0.0005306387  1.083078e-04  0.0005198178
#>  [4,] -0.001239560 -0.0006152904 -0.0003019923  6.743838e-05 -0.0006087288
#>  [5,]  0.003375873  0.0021344614  0.0012853315  1.233314e-04  0.0016261419
#>  [6,]  0.001162541  0.0008668530  0.0005831461  1.279167e-04  0.0005695098
#>  [7,]  0.001250155  0.0009610138  0.0006674281  1.515745e-04  0.0006320926
#>  [8,] -0.001209734 -0.0007855717 -0.0005284741 -4.230765e-05 -0.0006601845
#>  [9,]  0.003165119  0.0021492929  0.0013862138  1.963465e-04  0.0015819668
#> [10,]  0.002115437  0.0014403229  0.0009211800  1.465672e-04  0.0010409930
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0014370540  0.0011888026  1.018912e-03  0.0007006015  4.074368e-04
#>  [2,]  0.0010039544  0.0007936804  6.440612e-04  0.0003910772  1.429621e-04
#>  [3,]  0.0005776745  0.0005434299  5.037138e-04  0.0003867720  3.153396e-04
#>  [4,] -0.0004659381 -0.0002140423 -9.305767e-06  0.0001777771  4.515264e-04
#>  [5,]  0.0015850627  0.0012108879  9.416458e-04  0.0005271598  9.318451e-05
#>  [6,]  0.0006446753  0.0006001924  5.659396e-04  0.0004444259  3.740291e-04
#>  [7,]  0.0007251628  0.0006968276  6.679606e-04  0.0005296015  4.621714e-04
#>  [8,] -0.0006262597 -0.0004938645 -3.799713e-04 -0.0002062290 -4.057082e-05
#>  [9,]  0.0016040945  0.0013433760  1.163059e-03  0.0007874391  4.309969e-04
#> [10,]  0.0010854441  0.0009181877  8.032526e-04  0.0005605690  3.402485e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  3.943311e-04  3.510104e-04  2.325583e-04  1.229958e-04  2.365086e-04
#>  [2,]  1.593657e-04  1.699873e-04  9.955720e-05  4.187832e-05  1.342573e-04
#>  [3,]  2.898325e-04  2.117681e-04  1.514786e-04  8.993750e-05  1.206126e-04
#>  [4,]  3.487281e-04  1.572670e-04  1.306694e-04  1.046287e-04  2.528785e-05
#>  [5,]  1.465411e-04  2.163599e-04  1.170805e-04  3.907267e-05  1.902714e-04
#>  [6,]  3.378448e-04  2.471783e-04  1.738871e-04  1.047160e-04  1.380043e-04
#>  [7,]  4.194854e-04  2.978662e-04  2.098141e-04  1.266388e-04  1.642746e-04
#>  [8,] -6.881416e-05 -8.801147e-05 -4.585433e-05 -1.026635e-05 -8.201303e-05
#>  [9,]  4.278572e-04  3.760567e-04  2.441305e-04  1.266858e-04  2.572574e-04
#> [10,]  3.277659e-04  2.804873e-04  1.825627e-04  9.993259e-05  1.818609e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  8.440175e-05  5.869375e-05  3.352298e-05  4.136045e-05  3.111599e-05
#>  [2,]  5.133012e-05  2.825458e-05  1.364806e-05  2.707718e-05  1.668924e-05
#>  [3,]  4.384565e-05  3.752250e-05  2.594247e-05  1.980919e-05  1.919118e-05
#>  [4,]  1.082681e-05  2.872157e-05  3.053228e-05  1.596342e-06  1.268070e-05
#>  [5,]  7.294737e-05  3.551396e-05  1.233793e-05  3.872511e-05  2.052630e-05
#>  [6,]  5.111385e-05  4.310739e-05  2.991949e-05  2.378573e-05  2.219644e-05
#>  [7,]  6.014898e-05  5.227587e-05  3.725038e-05  2.747827e-05  2.633366e-05
#>  [8,] -2.708179e-05 -1.190926e-05 -4.660354e-06 -1.641147e-05 -8.340516e-06
#>  [9,]  9.303494e-05  6.301371e-05  3.619714e-05  4.575235e-05  3.395126e-05
#> [10,]  6.737818e-05  4.732526e-05  2.772952e-05  3.273032e-05  2.496105e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  3.300304e-05  3.206995e-05  3.438805e-05  2.461372e-05  2.988861e-05
#>  [2,]  2.191238e-05  2.066983e-05  2.452434e-05  1.289040e-05  1.861675e-05
#>  [3,]  1.627914e-05  1.630331e-05  1.515822e-05  1.479464e-05  1.504828e-05
#>  [4,] -1.658588e-06  1.822090e-06 -7.203614e-06  1.036689e-05  9.755379e-07
#>  [5,]  3.147312e-05  2.937406e-05  3.727845e-05  1.586307e-05  2.657904e-05
#>  [6,]  1.806496e-05  1.903417e-05  1.723803e-05  1.749285e-05  1.733014e-05
#>  [7,]  2.137612e-05  2.184255e-05  1.965184e-05  2.078169e-05  2.013159e-05
#>  [8,] -1.403458e-05 -1.135356e-05 -1.376523e-05 -6.233306e-06 -1.032841e-05
#>  [9,]  3.601299e-05  3.486611e-05  3.816037e-05  2.700713e-05  3.253188e-05
#> [10,]  2.552552e-05  2.554932e-05  2.689198e-05  1.963298e-05  2.322059e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  4.204839e-04  3.659200e-05  0.0006331387  4.850974e-04  4.254580e-04
#>  [2,]  2.641721e-04  2.459240e-05  0.0004274452  2.872275e-04  1.972874e-04
#>  [3,]  1.935457e-04  1.521798e-05  0.0002627468  2.677291e-04  3.004799e-04
#>  [4,] -2.069867e-05 -7.712752e-06 -0.0001456671  5.613499e-05  2.561633e-04
#>  [5,]  3.900398e-04  3.832351e-05  0.0006637214  4.031329e-04  2.184630e-04
#>  [6,]  2.221356e-04  1.775222e-05  0.0002988187  3.007990e-04  3.417564e-04
#>  [7,]  2.586759e-04  2.005706e-05  0.0003412826  3.522942e-04  4.095265e-04
#>  [8,] -1.583628e-04 -1.413469e-05 -0.0002594148 -1.459145e-04 -8.379458e-05
#>  [9,]  4.622627e-04  3.996745e-05  0.0006938290  5.280347e-04  4.525085e-04
#> [10,]  3.198595e-04  2.804872e-05  0.0004823767  3.859862e-04  3.484163e-04
#> 
#> $vW[[2]]
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,]  0.1648542  0.1218011  0.2068076  0.1543055  0.1444333  0.1832264
#> [2,] -0.1648542 -0.1218011 -0.2068076 -0.1543055 -0.1444333 -0.1832264
#>            [,7]        [,8]       [,9]    [,10]
#> [1,]  0.1876254  0.09635229  0.1578276  0.16705
#> [2,] -0.1876254 -0.09635229 -0.1578276 -0.16705
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5398128 0.2545529 0.7881142 0.2223330 0.1757085 0.5083607 0.4233932
#>  [8] 0.1831786 0.4256778 0.2073835
#> 
#> $B[[2]]
#> [1] -0.02125828  0.07082501
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.005520272  0.003994170  0.002165304 -0.001787940  0.006244831
#>  [6]  0.002436434  0.002751605 -0.002503827  0.006139454  0.004204539
#> 
#> $vB[[2]]
#> [1]  0.2271228 -0.2271228
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#>  [2,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [3,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#>  [4,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#>  [5,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#>  [6,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#>  [7,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#>  [8,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#>  [9,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [10,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [11,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [12,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [13,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [14,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [15,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [16,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [17,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [18,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [19,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [20,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [21,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [22,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [23,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [24,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [25,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [26,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [27,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [28,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [29,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [30,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [31,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [32,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [33,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [34,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [35,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [36,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [37,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [38,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [39,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [40,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#>  [2,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [3,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#>  [4,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#>  [5,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#>  [6,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#>  [7,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#>  [8,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#>  [9,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [10,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [11,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [12,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [13,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [14,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [15,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [16,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [17,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [18,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [19,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [20,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [21,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [22,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [23,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [24,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [25,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [26,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [27,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [28,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [29,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [30,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [31,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [32,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [33,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [34,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [35,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [36,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [37,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [38,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [39,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [40,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#>  [2,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [3,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#>  [4,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#>  [5,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#>  [6,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#>  [7,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#>  [8,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#>  [9,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [10,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [11,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [12,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [13,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [14,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [15,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [16,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [17,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [18,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [19,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [20,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [21,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [22,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [23,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [24,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [25,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [26,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [27,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [28,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [29,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [30,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [31,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [32,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [33,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [34,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [35,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [36,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [37,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [38,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [39,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [40,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#>  [2,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [3,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#>  [4,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#>  [5,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#>  [6,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#>  [7,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#>  [8,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#>  [9,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [10,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [11,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [12,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [13,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [14,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [15,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [16,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [17,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [18,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [19,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [20,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [21,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [22,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [23,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [24,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [25,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [26,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [27,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [28,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [29,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [30,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [31,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [32,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [33,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [34,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [35,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [36,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [37,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [38,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [39,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [40,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#>  [2,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [3,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#>  [4,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#>  [5,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#>  [6,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#>  [7,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#>  [8,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#>  [9,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [10,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [11,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [12,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [13,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [14,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [15,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [16,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [17,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [18,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [19,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [20,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [21,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [22,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [23,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [24,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [25,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [26,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [27,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [28,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [29,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [30,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [31,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [32,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [33,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [34,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [35,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [36,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [37,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [38,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [39,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [40,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#>  [2,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [3,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#>  [4,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#>  [5,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#>  [6,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#>  [7,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#>  [8,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#>  [9,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [10,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [11,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [12,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [13,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [14,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [15,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [16,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [17,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [18,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [19,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [20,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [21,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [22,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [23,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [24,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [25,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [26,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [27,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [28,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [29,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [30,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [31,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [32,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [33,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [34,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [35,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [36,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [37,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [38,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [39,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [40,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6204030 0.3971732 0.8753787 0.5578935 0.4970278 0.6856720 0.7699979
#>  [2,] 0.7443898 0.4981011 0.9386139 0.7300719 0.5828632 0.7908140 0.8635541
#>  [3,] 0.6906874 0.5024615 0.8909797 0.6297976 0.6014960 0.7700898 0.7981775
#>  [4,] 0.7801717 0.5430820 0.9403585 0.7545631 0.6267004 0.8239596 0.8805969
#>  [5,] 0.7195088 0.4668658 0.9157060 0.6324051 0.5632306 0.7948745 0.8271961
#>  [6,] 0.6951479 0.4082922 0.9055844 0.6240646 0.5186812 0.7524918 0.7980098
#>  [7,] 0.7173000 0.4239369 0.9131762 0.6215979 0.5222326 0.7870140 0.8082871
#>  [8,] 0.6520982 0.4302331 0.8957196 0.6361643 0.5148058 0.7252758 0.7899747
#>  [9,] 0.6144365 0.3893093 0.8257356 0.4425388 0.4212823 0.6532774 0.7166256
#> [10,] 0.7133911 0.5369049 0.8826948 0.6322960 0.5845614 0.7645292 0.7889637
#> [11,] 0.7213594 0.5293293 0.8877521 0.6337906 0.5780766 0.7710248 0.7938556
#> [12,] 0.7019854 0.4595747 0.9279581 0.7034723 0.5479979 0.7827528 0.8443194
#> [13,] 0.7020971 0.4287856 0.8848040 0.5383313 0.4963849 0.7481503 0.7756266
#> [14,] 0.7714575 0.4898011 0.9567369 0.7962599 0.6319757 0.8389713 0.8894819
#> [15,] 0.7325531 0.5343135 0.9134473 0.6707672 0.6255434 0.8108279 0.8397808
#> [16,] 0.6424409 0.3805230 0.8761483 0.5394297 0.4855195 0.6963788 0.7367772
#> [17,] 0.7651699 0.5855819 0.9241703 0.7152548 0.6350221 0.8144216 0.8542283
#> [18,] 0.7396084 0.5289474 0.9264140 0.6963116 0.6277790 0.8033539 0.8471752
#> [19,] 0.6947981 0.5044058 0.8694082 0.5814161 0.5892495 0.7812721 0.7788800
#> [20,] 0.7734182 0.5632003 0.9401646 0.7407288 0.6741912 0.8479747 0.8811452
#> [21,] 0.7099716 0.4268224 0.9059313 0.6459909 0.5444045 0.7498825 0.8072719
#> [22,] 0.7957988 0.6095362 0.9340192 0.7679729 0.6548175 0.8254319 0.8712817
#> [23,] 0.6829098 0.3479491 0.9025118 0.5136405 0.4362001 0.7624402 0.7903294
#> [24,] 0.7088102 0.4370457 0.9291109 0.6905720 0.5415882 0.7713089 0.8409991
#> [25,] 0.5969655 0.3271310 0.8667729 0.5096630 0.4303055 0.6725855 0.7138162
#> [26,] 0.6832103 0.5078768 0.8818614 0.6171164 0.6044852 0.7824931 0.7981747
#> [27,] 0.7066965 0.4294188 0.9142044 0.6048668 0.5061318 0.7447266 0.8361686
#> [28,] 0.7281971 0.5683463 0.8751046 0.6686120 0.6070081 0.7582982 0.7908883
#> [29,] 0.6518629 0.4414529 0.8977972 0.6162710 0.5288185 0.7152456 0.7969419
#> [30,] 0.7669954 0.5130628 0.9485348 0.7501693 0.5970937 0.8098691 0.8808346
#> [31,] 0.7305082 0.4365705 0.9217746 0.6586568 0.5452610 0.7924698 0.8267829
#> [32,] 0.7643182 0.5639062 0.9192165 0.6978098 0.6372280 0.8279211 0.8489404
#> [33,] 0.7709407 0.4898459 0.9540622 0.7848877 0.6261425 0.8347212 0.8842356
#> [34,] 0.6774060 0.3703398 0.9038143 0.6245576 0.4865330 0.7267985 0.7882207
#> [35,] 0.6908983 0.4251707 0.9057373 0.5800701 0.5049180 0.7334475 0.8223074
#> [36,] 0.6202158 0.3910443 0.8785275 0.5868675 0.4785793 0.6753773 0.7763137
#> [37,] 0.7276510 0.4522237 0.9340876 0.7098789 0.5445701 0.7665535 0.8553620
#> [38,] 0.6641375 0.4398399 0.9222519 0.5878752 0.5835068 0.8050549 0.8350066
#> [39,] 0.6292676 0.4075550 0.8210811 0.4840428 0.4675167 0.6881542 0.6873395
#> [40,] 0.7379103 0.5151724 0.9176969 0.6764051 0.5862116 0.7851899 0.8292916
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3204970 0.6086110 0.6573846
#>  [2,] 0.3480001 0.7665485 0.7177685
#>  [3,] 0.4123919 0.6574658 0.7074534
#>  [4,] 0.4164910 0.7905411 0.7376730
#>  [5,] 0.3625926 0.6586414 0.7039220
#>  [6,] 0.3120879 0.6526879 0.6850359
#>  [7,] 0.3344397 0.6428807 0.6969470
#>  [8,] 0.2951507 0.6703767 0.6404492
#>  [9,] 0.2727995 0.4904000 0.5915214
#> [10,] 0.4261107 0.6777339 0.6958340
#> [11,] 0.4084992 0.6773175 0.7006442
#> [12,] 0.3163940 0.7641573 0.6927046
#> [13,] 0.3428291 0.5676647 0.6768562
#> [14,] 0.3667140 0.8077346 0.7567749
#> [15,] 0.4099605 0.6878184 0.7488167
#> [16,] 0.2956689 0.5618767 0.6498137
#> [17,] 0.4387517 0.7442841 0.7309928
#> [18,] 0.3749266 0.7318155 0.7308391
#> [19,] 0.3846659 0.5826620 0.6997891
#> [20,] 0.4445008 0.7540551 0.7799611
#> [21,] 0.3691503 0.7122828 0.6759549
#> [22,] 0.4656688 0.7837131 0.7532472
#> [23,] 0.2903120 0.5470705 0.6656644
#> [24,] 0.3051230 0.7406820 0.6879338
#> [25,] 0.2707372 0.5866469 0.5882691
#> [26,] 0.3694399 0.6152347 0.7238095
#> [27,] 0.3022300 0.6769922 0.6884819
#> [28,] 0.4744395 0.7076364 0.6716272
#> [29,] 0.2950023 0.6633105 0.6584409
#> [30,] 0.3516546 0.7903724 0.7416463
#> [31,] 0.3747786 0.6821689 0.6829704
#> [32,] 0.4440096 0.7214842 0.7433969
#> [33,] 0.3713893 0.7949341 0.7559423
#> [34,] 0.2785608 0.6630110 0.6419175
#> [35,] 0.2870404 0.6414848 0.6884578
#> [36,] 0.2619537 0.6283638 0.6130529
#> [37,] 0.3261292 0.7645033 0.6917178
#> [38,] 0.3009855 0.6128993 0.7382627
#> [39,] 0.2970634 0.5001812 0.6152537
#> [40,] 0.3960385 0.7365356 0.6884135
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6789243 0.3210757
#>  [2,] 0.6917080 0.3082920
#>  [3,] 0.6864632 0.3135368
#>  [4,] 0.6947453 0.3052547
#>  [5,] 0.6873132 0.3126868
#>  [6,] 0.6839838 0.3160162
#>  [7,] 0.6852032 0.3147968
#>  [8,] 0.6823409 0.3176591
#>  [9,] 0.6723779 0.3276221
#> [10,] 0.6868585 0.3131415
#> [11,] 0.6872520 0.3127480
#> [12,] 0.6889508 0.3110492
#> [13,] 0.6814565 0.3185435
#> [14,] 0.6953486 0.3046514
#> [15,] 0.6909058 0.3090942
#> [16,] 0.6777443 0.3222557
#> [17,] 0.6933502 0.3066498
#> [18,] 0.6921816 0.3078184
#> [19,] 0.6845313 0.3154687
#> [20,] 0.6958971 0.3041029
#> [21,] 0.6858055 0.3141945
#> [22,] 0.6958964 0.3041036
#> [23,] 0.6792343 0.3207657
#> [24,] 0.6880901 0.3119099
#> [25,] 0.6739664 0.3260336
#> [26,] 0.6861969 0.3138031
#> [27,] 0.6858106 0.3141894
#> [28,] 0.6874418 0.3125582
#> [29,] 0.6831906 0.3168094
#> [30,] 0.6941215 0.3058785
#> [31,] 0.6866429 0.3133571
#> [32,] 0.6928336 0.3071664
#> [33,] 0.6947794 0.3052206
#> [34,] 0.6816993 0.3183007
#> [35,] 0.6844171 0.3155829
#> [36,] 0.6784702 0.3215298
#> [37,] 0.6892842 0.3107158
#> [38,] 0.6870163 0.3129837
#> [39,] 0.6739451 0.3260549
#> [40,] 0.6897711 0.3102289
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]         [,2]     [,3]        [,4]        [,5]      [,6]
#>  [1,] 0.4912593 -0.417257476 1.949377  0.23261713 -0.01188903 0.7799622
#>  [2,] 1.0689110 -0.007595714 2.727221  0.99498722  0.33453827 1.3298389
#>  [3,] 0.8033351  0.009845920 2.100788  0.53134867  0.41170235 1.2088182
#>  [4,] 1.2666675  0.172756364 2.757910  1.12309884  0.51808725 1.5434078
#>  [5,] 0.9420262 -0.132731462 2.385385  0.54254870  0.25428381 1.3545621
#>  [6,] 0.8242979 -0.371029963 2.260874  0.50683674  0.07475953 1.1119463
#>  [7,] 0.9311077 -0.306632595 2.353048  0.49633587  0.08898904 1.3070195
#>  [8,] 0.6282748 -0.280900268 2.150544  0.55875443  0.05924033 0.9707844
#>  [9,] 0.4659999 -0.450216335 1.555701 -0.23086468 -0.31751156 0.6334770
#> [10,] 0.9119113  0.147888709 2.018200  0.54207940  0.34152709 1.1776735
#> [11,] 0.9512147  0.117452178 2.067982  0.54851332  0.31488270 1.2141069
#> [12,] 0.8567699 -0.162055012 2.555739  0.86388794  0.19258463 1.2817813
#> [13,] 0.8573040 -0.286807794 2.038731  0.15362657 -0.01446072 1.0887716
#> [14,] 1.2165592 -0.040801402 3.096229  1.36308036  0.54070177 1.6505937
#> [15,] 1.0076150  0.137470283 2.356472  0.71165684  0.51314477 1.4553988
#> [16,] 0.5859741 -0.487328885 1.956450  0.15804692 -0.05793810 0.8301129
#> [17,] 1.1812356  0.345730653 2.500407  0.92104404  0.55382298 1.4790008
#> [18,] 1.0439343  0.115919365 2.532867  0.82979510  0.52270040 1.4073898
#> [19,] 0.8226478  0.017623666 1.895737  0.32858916  0.36086361 1.2730951
#> [20,] 1.2277139  0.254160801 2.754458  1.04975992  0.72720305 1.7188038
#> [21,] 0.8952461 -0.294827442 2.264938  0.60146283  0.17808723 1.0979858
#> [22,] 1.3602408  0.445362988 2.650133  1.19690042  0.64028312 1.5535916
#> [23,] 0.7671763 -0.628066427 2.225450  0.05457566 -0.25659834 1.1661047
#> [24,] 0.8896124 -0.253160598 2.573112  0.80279476  0.16673798 1.2157165
#> [25,] 0.3928373 -0.721190025 1.872721  0.03865695 -0.28060500 0.7199022
#> [26,] 0.7685646  0.031509835 2.010177  0.47732652  0.42418903 1.2802546
#> [27,] 0.8793934 -0.284222744 2.366087  0.42578552  0.02452838 1.0706821
#> [28,] 0.9854943  0.275107159 1.946867  0.70191407  0.43475310 1.1433721
#> [29,] 0.6272379 -0.235267475 2.172985  0.47375033  0.11540179 0.9209991
#> [30,] 1.1914225  0.052263084 2.914012  1.09951538  0.39337019 1.4491597
#> [31,] 0.9972026 -0.255092360 2.466706  0.65731403  0.18154095 1.3398775
#> [32,] 1.1765016  0.257030703 2.431749  0.83689019  0.56335312 1.5709648
#> [33,] 1.2136307 -0.040621908 3.033441  1.29438032  0.51570336 1.6194641
#> [34,] 0.7418761 -0.530759405 2.240343  0.50893837 -0.05388104 0.9784398
#> [35,] 0.8043223 -0.301582312 2.262663  0.32306134  0.01967277 1.0121849
#> [36,] 0.4904645 -0.442924525 1.978560  0.35103061 -0.08573512 0.7326081
#> [37,] 0.9827373 -0.191690077 2.651244  0.89479620  0.17875479 1.1889518
#> [38,] 0.6817873 -0.241811936 2.473344  0.35518849  0.33718613 1.4181925
#> [39,] 0.5290762 -0.374082237 1.523689 -0.06385063 -0.13011641 0.7915040
#> [40,] 1.0351357  0.060708185 2.411458  0.73729964  0.34832597 1.2961713
#>            [,7]       [,8]          [,9]     [,10]
#>  [1,] 1.2082993 -0.7514886  0.4414772823 0.6516607
#>  [2,] 1.8451286 -0.6278416  1.1889234567 0.9334195
#>  [3,] 1.3749426 -0.3540860  0.6520211407 0.8830476
#>  [4,] 1.9980945 -0.3371948  1.3281902884 1.0339087
#>  [5,] 1.5658842 -0.5641292  0.6572457900 0.8660446
#>  [6,] 1.3739017 -0.7903762  0.6308753333 0.7770124
#>  [7,] 1.4389182 -0.6881728  0.5878891735 0.8328015
#>  [8,] 1.3247729 -0.8704978  0.7098893645 0.5773142
#>  [9,] 0.9277846 -0.9804654 -0.0384045783 0.3702582
#> [10,] 1.3186902 -0.2977373  0.7433772253 0.8275377
#> [11,] 1.3483244 -0.3701730  0.7414714006 0.8503673
#> [12,] 1.6907245 -0.7703934  1.1756086934 0.8127942
#> [13,] 1.2403596 -0.6507120  0.2723294690 0.7393615
#> [14,] 2.0854600 -0.5463397  1.4353566237 1.1350787
#> [15,] 1.6565981 -0.3641287  0.7899399728 1.0923115
#> [16,] 1.0292848 -0.8680081  0.2487821234 0.6182205
#> [17,] 1.7681566 -0.2462296  1.0683559532 0.9996657
#> [18,] 1.7126155 -0.5111387  1.0038531633 0.9988839
#> [19,] 1.2591514 -0.4697903  0.3337108105 0.8462939
#> [20,] 2.0033202 -0.2229153  1.1203577658 1.2654399
#> [21,] 1.4323801 -0.5358637  0.9064968611 0.7352435
#> [22,] 1.9123388 -0.1375411  1.2874369795 1.1160066
#> [23,] 1.3269121 -0.8938691  0.1888409997 0.6886403
#> [24,] 1.6656811 -0.8230200  1.0495164090 0.7904772
#> [25,] 0.9139911 -0.9908857  0.3501209949 0.3568145
#> [26,] 1.3749249 -0.5346205  0.4693703012 0.9634372
#> [27,] 1.6299926 -0.8367012  0.7399830650 0.7930318
#> [28,] 1.3302885 -0.1023312  0.8839320437 0.7155540
#> [29,] 1.3672894 -0.8712116  0.6780820497 0.6563540
#> [30,] 2.0003570 -0.6117740  1.3271717337 1.0545427
#> [31,] 1.5629965 -0.5117705  0.7637572726 0.7674565
#> [32,] 1.7263147 -0.2249046  0.9518358172 1.0636996
#> [33,] 2.0331666 -0.5262611  1.3549277326 1.1305604
#> [34,] 1.3142337 -0.9516118  0.6767411487 0.5836964
#> [35,] 1.5320590 -0.9098018  0.5818142746 0.7929194
#> [36,] 1.2443120 -1.0358386  0.5252039410 0.4601631
#> [37,] 1.7772903 -0.7257446  1.1775292270 0.8081625
#> [38,] 1.6215340 -0.8426096  0.4595159547 1.0369583
#> [39,] 0.7877104 -0.8613212  0.0007246389 0.4694504
#> [40,] 1.5806152 -0.4219992  1.0280392519 0.7927126
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3877876 -0.3610452
#>  [2,] 0.4208363 -0.3872804
#>  [3,] 0.4027154 -0.3809206
#>  [4,] 0.4241718 -0.3982268
#>  [5,] 0.4117512 -0.3758369
#>  [6,] 0.4065104 -0.3656303
#>  [7,] 0.4098774 -0.3679105
#>  [8,] 0.4016438 -0.3629069
#>  [9,] 0.3796725 -0.3392872
#> [10,] 0.4027419 -0.3827313
#> [11,] 0.4058318 -0.3814714
#> [12,] 0.4167456 -0.3784734
#> [13,] 0.3991264 -0.3613470
#> [14,] 0.4305873 -0.3946577
#> [15,] 0.4145018 -0.3898556
#> [16,] 0.3914173 -0.3520074
#> [17,] 0.4183048 -0.3975238
#> [18,] 0.4210457 -0.3892927
#> [19,] 0.4034861 -0.3711886
#> [20,] 0.4246296 -0.4032061
#> [21,] 0.4059154 -0.3746666
#> [22,] 0.4217779 -0.4060546
#> [23,] 0.3996416 -0.3506136
#> [24,] 0.4161804 -0.3750253
#> [25,] 0.3848553 -0.3413244
#> [26,] 0.4061826 -0.3762163
#> [27,] 0.4094948 -0.3711109
#> [28,] 0.4003693 -0.3878174
#> [29,] 0.4020280 -0.3664457
#> [30,] 0.4267897 -0.3926692
#> [31,] 0.4103338 -0.3741373
#> [32,] 0.4180030 -0.3953974
#> [33,] 0.4286669 -0.3938928
#> [34,] 0.4039751 -0.3576170
#> [35,] 0.4069253 -0.3672207
#> [36,] 0.3927941 -0.3539560
#> [37,] 0.4163363 -0.3804385
#> [38,] 0.4145117 -0.3716953
#> [39,] 0.3846616 -0.3414211
#> [40,] 0.4140757 -0.3849736
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.6789243  0.6789243
#>  [2,]  0.3082920 -0.3082920
#>  [3,] -0.6864632  0.6864632
#>  [4,]  0.3052547 -0.3052547
#>  [5,] -0.6873132  0.6873132
#>  [6,]  0.3160162 -0.3160162
#>  [7,] -0.6852032  0.6852032
#>  [8,] -0.6823409  0.6823409
#>  [9,]  0.3276221 -0.3276221
#> [10,] -0.6868585  0.6868585
#> [11,] -0.6872520  0.6872520
#> [12,] -0.6889508  0.6889508
#> [13,] -0.6814565  0.6814565
#> [14,] -0.6953486  0.6953486
#> [15,] -0.6909058  0.6909058
#> [16,] -0.6777443  0.6777443
#> [17,]  0.3066498 -0.3066498
#> [18,] -0.6921816  0.6921816
#> [19,] -0.6845313  0.6845313
#> [20,] -0.6958971  0.6958971
#> [21,] -0.6858055  0.6858055
#> [22,] -0.6958964  0.6958964
#> [23,]  0.3207657 -0.3207657
#> [24,]  0.3119099 -0.3119099
#> [25,]  0.3260336 -0.3260336
#> [26,] -0.6861969  0.6861969
#> [27,]  0.3141894 -0.3141894
#> [28,] -0.6874418  0.6874418
#> [29,] -0.6831906  0.6831906
#> [30,]  0.3058785 -0.3058785
#> [31,]  0.3133571 -0.3133571
#> [32,] -0.6928336  0.6928336
#> [33,] -0.6947794  0.6947794
#> [34,]  0.3183007 -0.3183007
#> [35,]  0.3155829 -0.3155829
#> [36,] -0.6784702  0.6784702
#> [37,]  0.3107158 -0.3107158
#> [38,] -0.6870163  0.6870163
#> [39,]  0.3260549 -0.3260549
#> [40,] -0.6897711  0.6897711
#> 
#> $L
#> [1] 0.6955530 0.6839152 0.7085716 0.7048702 0.6697223 0.8686338
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5072464