Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13          V14
#>  [1,] -0.2096490 -0.04342449 -0.114921641 -0.064985152 -0.01538859 -0.148794442
#>  [2,] -0.3259202 -0.14844087 -0.021593221 -0.015924433 -0.10423272 -0.074163705
#>  [3,] -0.2394809 -0.05943662  0.009585767 -0.067471343 -0.07969321 -0.157129172
#>  [4,] -0.3374681 -0.17101417 -0.195561669 -0.180825510 -0.12385524 -0.067179412
#>  [5,] -0.2900300 -0.21151325 -0.157717288 -0.145449946 -0.08692028 -0.062748237
#>  [6,] -0.3683469 -0.06129415  0.026267091  0.005283862  0.05477124  0.031218828
#>  [7,] -0.3272190 -0.18127563 -0.131676882 -0.026185463 -0.18288935 -0.006928521
#>  [8,] -0.3800435 -0.02086998 -0.123377570 -0.060026815 -0.05994930  0.061567662
#>  [9,] -0.2159846 -0.16790199 -0.061511674 -0.193609898 -0.13165822 -0.077441132
#> [10,] -0.2843309 -0.07528669 -0.161496645 -0.024427122 -0.14982489 -0.014536126
#>                V15          V16         V17         V18           V19
#>  [1,]  0.008152452  0.056872802 -0.01258248  0.03344207  0.0167451929
#>  [2,]  0.073446394  0.094699479  0.03009439  0.02896326 -0.0305366049
#>  [3,] -0.061542827 -0.051693932 -0.06249256 -0.14147534  0.0513890512
#>  [4,] -0.085799626  0.019316942 -0.04297680 -0.11042900 -0.0186448663
#>  [5,] -0.006172962 -0.018800704 -0.04930822  0.09874620  0.1185777634
#>  [6,]  0.099927087 -0.003102435  0.17656362  0.17702066  0.1282666482
#>  [7,] -0.170716242  0.043365394 -0.01233165 -0.03709833  0.0238982009
#>  [8,] -0.016729341  0.039907189  0.07362723  0.18206407  0.0508250753
#>  [9,]  0.010662672 -0.096017292 -0.01974467 -0.01469723  0.0340497322
#> [10,] -0.069553360 -0.049550561 -0.04159191 -0.08676390 -0.0001542875
#>               V2         V20         V21         V22         V23          V24
#>  [1,] -0.3216275  0.19735571  0.06816959  0.02749662  0.20956225  0.103478254
#>  [2,] -0.3736825  0.06523310 -0.02866963  0.11465108  0.16097152  0.208653446
#>  [3,] -0.2062435  0.01553328  0.08881198  0.01303090 -0.01068513 -0.007037518
#>  [4,] -0.2833909 -0.05582291  0.10519660  0.08605692  0.01523502  0.114513941
#>  [5,] -0.3184150  0.04468877  0.14564675  0.09339704  0.11609833  0.069225473
#>  [6,] -0.3139621  0.12224305  0.16701126  0.14748270  0.19951589  0.125727362
#>  [7,] -0.2262038 -0.04923981  0.09307845 -0.08502295  0.07009604  0.127632943
#>  [8,] -0.3312473  0.11999389  0.17475246  0.19380966  0.24456308  0.250280722
#>  [9,] -0.2493851 -0.04946655  0.07388557  0.04911882  0.06737579  0.072425832
#> [10,] -0.2160316  0.04254395  0.05102961  0.11206854  0.05378552  0.134402524
#>              V25         V26        V27        V28        V29         V3
#>  [1,] 0.21665747  0.16118314 0.22304899 0.09190069 0.05151299 -0.2329530
#>  [2,] 0.02871486 -0.02463657 0.06358802 0.01697457 0.04647839 -0.4029174
#>  [3,] 0.02283561  0.03510955 0.08376023 0.09114887 0.16028827 -0.2580323
#>  [4,] 0.17637422  0.17233988 0.13848735 0.08232554 0.15447481 -0.2763329
#>  [5,] 0.18540981  0.09248403 0.10884721 0.27022772 0.11484993 -0.2967490
#>  [6,] 0.22116269  0.14711054 0.21209746 0.27865745 0.12025457 -0.2199368
#>  [7,] 0.03878034  0.18506139 0.02297505 0.10368697 0.03513400 -0.1914384
#>  [8,] 0.24578591  0.35053708 0.40777250 0.38346940 0.18233816 -0.3643885
#>  [9,] 0.18641328  0.10630175 0.08485074 0.11075206 0.19650860 -0.2769325
#> [10,] 0.06413714  0.21446556 0.19473880 0.01567502 0.15820328 -0.2019814
#>               V30         V31          V32          V33          V34
#>  [1,]  0.06098803 -0.03464060  0.052770120 -0.015119839 -0.042535182
#>  [2,]  0.04563819  0.01932370  0.064201575  0.031579075  0.017491007
#>  [3,]  0.03617831  0.01100677  0.058754177 -0.051728625 -0.016248750
#>  [4,]  0.01526818 -0.03236276 -0.084057702  0.015306215 -0.076929771
#>  [5,]  0.20183388  0.13682773  0.115688011  0.071046480  0.009168883
#>  [6,]  0.11129265  0.03431178  0.002330936  0.008745546  0.026635728
#>  [7,]  0.08735302 -0.02869237  0.021181398 -0.089632589  0.055008312
#>  [8,]  0.06641716  0.15577743 -0.021615607  0.117309238  0.025589560
#>  [9,]  0.15400314  0.07009800 -0.024104536  0.043974226  0.001993042
#> [10,] -0.05843158  0.07108108 -0.044943135 -0.084846238  0.004558913
#>                V35          V36          V37          V38         V39
#>  [1,]  0.004657964  0.052047803 -0.023225668 -0.078586488 -0.03804345
#>  [2,]  0.013680335  0.040448876 -0.030697129 -0.021411000 -0.14639790
#>  [3,] -0.142585077 -0.065431089 -0.137322958  0.014246790 -0.19369694
#>  [4,] -0.056421548 -0.137335688 -0.173787344 -0.141297824 -0.07818206
#>  [5,]  0.001377393  0.006918404 -0.128221310 -0.003079658 -0.08155931
#>  [6,] -0.007975122 -0.034974968  0.105131436 -0.010046835  0.02571420
#>  [7,]  0.019371926 -0.021629068 -0.041473481 -0.128363503 -0.09318895
#>  [8,]  0.058934024  0.016281471  0.062730344  0.015105514  0.01440443
#>  [9,]  0.035188770 -0.095492986 -0.074670078 -0.041878930 -0.05913803
#> [10,] -0.135225271 -0.106494670 -0.007613405 -0.174045314 -0.21886129
#>               V4           V40         V41           V42          V43
#>  [1,] -0.2701071 -0.0839671793  0.06151131  0.0001655821 -0.004627494
#>  [2,] -0.3596055 -0.0502350952 -0.13021534 -0.0896821227 -0.108630214
#>  [3,] -0.2840311 -0.1589024669 -0.18188841 -0.1115078465 -0.095038625
#>  [4,] -0.2859199 -0.0328903332 -0.05994539 -0.0936658082 -0.054081396
#>  [5,] -0.3044849 -0.0836695948  0.02482525 -0.0669913554 -0.070809776
#>  [6,] -0.2607140 -0.0966787805 -0.06553789 -0.0478521433 -0.101091537
#>  [7,] -0.1914358 -0.0337109893 -0.09059237 -0.0267418626 -0.033033551
#>  [8,] -0.1748404 -0.0005024087 -0.10392140 -0.0851302800 -0.066334839
#>  [9,] -0.1933334 -0.1637613497 -0.07423138  0.0124831061 -0.096266568
#> [10,] -0.3202792 -0.0371068876 -0.03561131 -0.1845698674 -0.218906931
#>                V44         V45         V46        V47        V48        V49
#>  [1,] -0.153058730 -0.04029178 -0.15873020 -0.1502458 -0.1684635 -0.2080194
#>  [2,] -0.071234580 -0.20242981 -0.16623812 -0.1731032 -0.2344660 -0.3284947
#>  [3,] -0.192595983 -0.07362957 -0.11591290 -0.2769361 -0.2841883 -0.3466805
#>  [4,] -0.093787421 -0.08488149 -0.20952073 -0.2421984 -0.1381760 -0.2043848
#>  [5,] -0.098393035 -0.14142411 -0.20782203 -0.1302216 -0.2331358 -0.2165498
#>  [6,] -0.139856996 -0.18198018 -0.21049738 -0.1489935 -0.2690355 -0.2003240
#>  [7,] -0.205393043 -0.12175307 -0.12144974 -0.1428423 -0.1086048 -0.2468007
#>  [8,] -0.008115959 -0.01287102 -0.05707213 -0.2287265 -0.1977827 -0.2308884
#>  [9,] -0.022362551 -0.20192524 -0.13983212 -0.1059997 -0.2744655 -0.2340052
#> [10,] -0.102396863 -0.19339252 -0.16869591 -0.1291835 -0.1408046 -0.3431388
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.2149360 -0.2192876 -0.3168733 -0.3913163 -0.2514449 -0.3132266
#>  [2,] -0.2309836 -0.3416065 -0.3886418 -0.2510643 -0.2501489 -0.3739906
#>  [3,] -0.2022291 -0.3649822 -0.2150230 -0.3655163 -0.3780146 -0.3941580
#>  [4,] -0.2600940 -0.2617032 -0.3936039 -0.3804612 -0.3983329 -0.3776087
#>  [5,] -0.3159529 -0.2671077 -0.4148205 -0.2410617 -0.3313354 -0.2525881
#>  [6,] -0.1493748 -0.3698496 -0.3279252 -0.4033123 -0.4159818 -0.2765623
#>  [7,] -0.2461493 -0.2163668 -0.3410785 -0.2869269 -0.3791468 -0.2971974
#>  [8,] -0.2085033 -0.3520885 -0.2682018 -0.3089767 -0.2657284 -0.3757040
#>  [9,] -0.1693283 -0.3094703 -0.2962829 -0.3406154 -0.3251487 -0.3577546
#> [10,] -0.2595179 -0.2641152 -0.3381720 -0.2752377 -0.4006559 -0.4027793
#>              V55        V56        V57        V58        V59         V6
#>  [1,] -0.3236644 -0.3162584 -0.3699220 -0.2875854 -0.3917274 -0.2449147
#>  [2,] -0.2841033 -0.4482921 -0.4198169 -0.3898309 -0.3121349 -0.1482447
#>  [3,] -0.3101320 -0.2378043 -0.3297497 -0.3745388 -0.3347215 -0.1058881
#>  [4,] -0.3188616 -0.3125619 -0.2995491 -0.2512809 -0.2586432 -0.1227998
#>  [5,] -0.3565610 -0.3437553 -0.4440312 -0.2784745 -0.4139176 -0.2604499
#>  [6,] -0.4078320 -0.3618851 -0.3526498 -0.4175460 -0.2888413 -0.2145909
#>  [7,] -0.2927516 -0.3442557 -0.2949778 -0.3032530 -0.3312964 -0.1593911
#>  [8,] -0.3854169 -0.3574638 -0.2915298 -0.3325757 -0.3243006 -0.1925978
#>  [9,] -0.2821043 -0.2770497 -0.2261345 -0.2665343 -0.3590214 -0.2671329
#> [10,] -0.3568508 -0.2519233 -0.3068195 -0.4169852 -0.3536840 -0.2099912
#>              V60         V7         V8          V9
#>  [1,] -0.3124941 -0.2216370 -0.1341838 -0.17879297
#>  [2,] -0.3372722 -0.2263527 -0.1270268 -0.21406186
#>  [3,] -0.3088094 -0.1063079 -0.1708200 -0.15025189
#>  [4,] -0.3304915 -0.1873495 -0.2171657 -0.14768837
#>  [5,] -0.3042475 -0.2522807 -0.1302497 -0.06911245
#>  [6,] -0.3654236 -0.1612253 -0.1204766 -0.13415938
#>  [7,] -0.2660640 -0.0980095 -0.1564237 -0.15911013
#>  [8,] -0.3144144 -0.2197429 -0.1982837 -0.17118811
#>  [9,] -0.3826415 -0.1101619 -0.2062068 -0.19618833
#> [10,] -0.2588910 -0.1399538 -0.2345136 -0.08725449
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.12215511 -0.11253132 -0.04385970 -0.09322432 0.009933229  0.03855021
#> [2,] -0.04199988  0.05266351 -0.02373333  0.02851139 0.029097620 -0.06647369
#>             [,7]         [,8]        [,9]      [,10]
#> [1,] -0.08978746 -0.009003081 -0.09676473 0.05226890
#> [2,]  0.08682296 -0.148697790  0.02169732 0.03033092
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.463159e-04 -1.416285e-03 -1.748552e-03 -1.733446e-03 -1.337418e-03
#>  [2,]  1.130291e-04  1.138104e-03  1.444145e-03  1.552956e-03  1.246384e-03
#>  [3,] -7.744591e-06 -9.012016e-05 -1.135463e-04 -9.673096e-05 -5.164217e-05
#>  [4,]  7.377106e-05  7.648576e-04  1.005458e-03  1.119706e-03  8.829093e-04
#>  [5,]  2.354506e-06  2.311877e-05  2.578963e-05  5.785114e-05  8.503162e-05
#>  [6,] -5.720737e-05 -5.340989e-04 -6.443533e-04 -5.801598e-04 -4.083764e-04
#>  [7,]  1.115425e-04  1.182802e-03  1.550034e-03  1.706042e-03  1.317722e-03
#>  [8,] -5.574887e-05 -4.603768e-04 -5.440410e-04 -4.658415e-04 -3.597581e-04
#>  [9,]  7.478522e-05  7.620531e-04  9.817237e-04  1.077591e-03  8.624505e-04
#> [10,] -5.116163e-05 -5.355663e-04 -6.898956e-04 -7.091795e-04 -5.272296e-04
#>                 V14           V15           V16           V17           V18
#>  [1,] -8.543280e-04 -2.875831e-04 -5.843795e-05 -1.321379e-04 -3.166532e-04
#>  [2,]  8.136916e-04  3.894312e-04  2.622330e-04  3.621963e-04  5.739640e-04
#>  [3,]  1.968994e-05  9.154066e-05  1.352693e-04  1.394069e-04  1.527570e-04
#>  [4,]  5.298912e-04  2.015279e-04  9.157095e-05  1.716236e-04  3.387604e-04
#>  [5,]  8.441086e-05  1.100035e-04  1.486932e-04  1.576691e-04  1.735864e-04
#>  [6,] -2.803025e-04 -7.846268e-05 -2.358448e-06 -1.953044e-05 -4.842706e-05
#>  [7,]  7.504501e-04  2.239131e-04  3.226440e-05  1.369762e-04  3.787096e-04
#>  [8,] -3.218433e-04 -1.788680e-04 -9.966829e-05 -1.124402e-04 -1.426058e-04
#>  [9,]  5.443470e-04  2.393392e-04  1.444686e-04  2.166116e-04  3.795059e-04
#> [10,] -2.317481e-04  5.506112e-05  1.827997e-04  1.491825e-04  9.239064e-05
#>                 V19            V2           V20           V21           V22
#>  [1,] -7.460644e-04 -1.219752e-04 -0.0010426001 -1.403693e-03 -1.086624e-03
#>  [2,]  8.828144e-04  1.125342e-04  0.0011480121  1.411779e-03  1.231294e-03
#>  [3,]  1.274096e-04 -3.272116e-06  0.0001256507  8.539637e-05  9.610949e-05
#>  [4,]  5.626942e-04  7.237082e-05  0.0007070238  8.935375e-04  8.021702e-04
#>  [5,]  1.651936e-04  1.513977e-05  0.0002369679  2.086322e-04  2.071493e-04
#>  [6,] -1.848193e-04 -3.644933e-05 -0.0002727201 -4.145986e-04 -2.515165e-04
#>  [7,]  7.182687e-04  1.058970e-04  0.0009256723  1.222991e-03  1.067950e-03
#>  [8,] -2.792198e-04 -3.569805e-05 -0.0003764691 -4.852141e-04 -3.029868e-04
#>  [9,]  5.996829e-04  7.575015e-05  0.0007765467  9.496879e-04  8.394444e-04
#> [10,] -7.050187e-05 -4.453509e-05 -0.0001652418 -3.263434e-04 -2.607881e-04
#>                 V23          V24          V25           V26           V27
#>  [1,] -7.099265e-04 3.319345e-05 3.498548e-05 -3.905637e-04 -4.418325e-04
#>  [2,]  1.052482e-03 5.632281e-04 5.322752e-04  7.990267e-04  8.573713e-04
#>  [3,]  1.467338e-04 2.060700e-04 2.118149e-04  1.905562e-04  1.768564e-04
#>  [4,]  6.818878e-04 3.089147e-04 2.705764e-04  4.408125e-04  4.842125e-04
#>  [5,]  2.566538e-04 3.102591e-04 3.097032e-04  2.924687e-04  2.961132e-04
#>  [6,] -7.609006e-05 2.509887e-04 2.299422e-04  4.211603e-05  2.507515e-05
#>  [7,]  8.567884e-04 2.722089e-04 2.089112e-04  4.664490e-04  5.295396e-04
#>  [8,] -9.535369e-05 2.500486e-04 2.253848e-04  3.203413e-05  4.685418e-05
#>  [9,]  7.240046e-04 3.762673e-04 3.493544e-04  5.220243e-04  5.582320e-04
#> [10,] -1.161376e-04 1.591714e-04 1.873004e-04  6.450455e-05  1.669448e-05
#>                 V28           V29            V3           V30           V31
#>  [1,] -2.965081e-04 -0.0001526588 -1.791877e-04 -3.629362e-04  0.0006265870
#>  [2,]  8.197814e-04  0.0007492575  1.648741e-04  9.044092e-04  0.0002169938
#>  [3,]  1.729642e-04  0.0001936596 -4.298332e-06  1.957736e-04  0.0002853378
#>  [4,]  4.481915e-04  0.0004151809  1.109104e-04  5.332626e-04  0.0000630717
#>  [5,]  3.199439e-04  0.0002954546  1.815917e-05  2.918598e-04  0.0003684552
#>  [6,]  1.074865e-04  0.0001562849 -6.003914e-05  4.665958e-05  0.0004707988
#>  [7,]  4.939238e-04  0.0004556120  1.621102e-04  6.480294e-04 -0.0001087308
#>  [8,]  1.692060e-04  0.0002340700 -5.677053e-05  1.212201e-04  0.0004884843
#>  [9,]  5.332670e-04  0.0004972648  1.118909e-04  6.175444e-04  0.0001625724
#> [10,]  3.224726e-05  0.0001054450 -6.429530e-05  5.871128e-05  0.0004340476
#>                 V32           V33           V34           V35           V36
#>  [1,]  5.887526e-04  5.731525e-04  0.0012769603  0.0019918344  0.0027182426
#>  [2,]  1.912831e-04  1.776481e-04 -0.0003560640 -0.0009959136 -0.0014678197
#>  [3,]  2.476792e-04  2.747424e-04  0.0003496910  0.0003733694  0.0004396456
#>  [4,]  4.607565e-05  3.641429e-05 -0.0003633726 -0.0008506639 -0.0011924668
#>  [5,]  3.889619e-04  3.766388e-04  0.0004148877  0.0004004373  0.0004605233
#>  [6,]  4.857391e-04  4.922479e-04  0.0007898940  0.0010625116  0.0013657734
#>  [7,] -1.322384e-04 -1.544946e-04 -0.0007897745 -0.0015223367 -0.0020655856
#>  [8,]  4.748500e-04  4.114266e-04  0.0006385007  0.0008543167  0.0011494638
#>  [9,]  1.512623e-04  1.507143e-04 -0.0002157187 -0.0006767505 -0.0009979457
#> [10,]  3.600724e-04  4.183609e-04  0.0007387520  0.0010124968  0.0012942060
#>                 V37           V38           V39            V4          V40
#>  [1,]  0.0019465198  4.203666e-04 -2.682360e-04 -4.723064e-04 0.0001098609
#>  [2,] -0.0009711940  4.649650e-05  5.973964e-04  3.742079e-04 0.0003568909
#>  [3,]  0.0003545544  1.650289e-04  1.044734e-04 -2.597443e-05 0.0001498022
#>  [4,] -0.0008196030 -5.130539e-05  3.879750e-04  2.601045e-04 0.0002288232
#>  [5,]  0.0003836690  2.711343e-04  2.279522e-04  2.176516e-06 0.0002320270
#>  [6,]  0.0009932667  3.703138e-04  9.588652e-05 -1.842390e-04 0.0002238079
#>  [7,] -0.0014607309 -2.193514e-04  4.738113e-04  3.968526e-04 0.0002267196
#>  [8,]  0.0008325858  2.946050e-04  1.004374e-04 -1.721886e-04 0.0002323390
#>  [9,] -0.0006677184  4.223806e-05  4.339333e-04  2.540160e-04 0.0002671298
#> [10,]  0.0009767229  2.669321e-04 -3.430734e-05 -1.726241e-04 0.0001432545
#>                 V41           V42           V43           V44           V45
#>  [1,] -1.971272e-04 -8.543025e-04 -1.157355e-03 -1.178578e-03 -1.594517e-03
#>  [2,]  5.827421e-04  1.012022e-03  1.126974e-03  1.079907e-03  1.327699e-03
#>  [3,]  1.173600e-04  3.643346e-05 -1.817635e-05 -5.384132e-05 -1.013516e-04
#>  [4,]  3.988565e-04  7.139497e-04  8.001064e-04  7.601484e-04  9.192017e-04
#>  [5,]  2.211901e-04  1.548250e-04  7.809965e-05  5.749401e-05  4.262425e-05
#>  [6,]  8.540679e-05 -1.796574e-04 -3.612231e-04 -3.838612e-04 -5.694897e-04
#>  [7,]  4.942033e-04  1.006615e-03  1.177438e-03  1.138051e-03  1.395776e-03
#>  [8,]  1.226863e-04 -1.111778e-04 -2.709481e-04 -2.610022e-04 -4.363248e-04
#>  [9,]  4.234772e-04  7.142318e-04  7.868394e-04  7.483096e-04  9.072547e-04
#> [10,]  3.857996e-06 -2.882449e-04 -4.211340e-04 -4.707948e-04 -6.417547e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -1.023716e-03 -4.774029e-04 -4.722066e-04 -2.794398e-04 -3.528235e-04
#>  [2,]  8.729351e-04  4.846148e-04  4.435549e-04  2.538443e-04  3.165328e-04
#>  [3,] -5.853654e-05 -1.510236e-05 -1.826464e-05 -1.304799e-05 -1.198981e-05
#>  [4,]  5.927158e-04  3.332694e-04  3.020316e-04  1.723772e-04  2.203039e-04
#>  [5,]  4.672842e-05  5.789680e-05  4.085241e-05  1.959906e-05  1.474133e-05
#>  [6,] -3.618152e-04 -1.367790e-04 -1.492739e-04 -9.142893e-05 -1.208765e-04
#>  [7,]  8.986805e-04  4.949321e-04  4.532106e-04  2.594005e-04  3.284223e-04
#>  [8,] -2.734723e-04 -7.938753e-05 -1.056194e-04 -6.685099e-05 -1.086091e-04
#>  [9,]  5.905635e-04  3.306736e-04  3.017816e-04  1.713016e-04  2.154756e-04
#> [10,] -4.031723e-04 -1.958204e-04 -1.857101e-04 -1.099001e-04 -1.281930e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -6.783021e-05 -9.746740e-05 -9.140056e-05 -2.517085e-05 -3.633748e-05
#>  [2,]  6.849811e-05  9.070176e-05  8.051046e-05  2.787989e-05  3.307221e-05
#>  [3,]  2.063803e-06 -2.814275e-06 -2.244375e-06  2.998770e-06  4.800493e-07
#>  [4,]  4.437161e-05  6.307845e-05  5.575659e-05  1.787114e-05  2.126378e-05
#>  [5,]  1.067495e-05  6.294243e-06  5.370811e-06  5.058014e-06  4.559855e-06
#>  [6,] -2.007463e-05 -3.127919e-05 -3.206407e-05 -6.502673e-06 -1.052996e-05
#>  [7,]  6.281180e-05  9.407392e-05  8.310791e-05  2.404970e-05  2.992469e-05
#>  [8,] -1.898349e-05 -2.581899e-05 -3.027778e-05 -9.393391e-06 -1.203820e-05
#>  [9,]  4.636057e-05  6.210242e-05  5.567786e-05  1.929979e-05  2.279598e-05
#> [10,] -1.966295e-05 -3.604919e-05 -3.172795e-05 -3.968959e-06 -1.055244e-05
#>                V55           V56           V57           V58           V59
#>  [1,] 2.790630e-06 -2.687456e-05 -2.760138e-06 -2.736038e-05 -1.035112e-05
#>  [2,] 7.541680e-06  2.952523e-05  1.004937e-05  2.527167e-05  1.577462e-05
#>  [3,] 4.421588e-06  1.592122e-06  3.231965e-06  3.374858e-07  1.937140e-06
#>  [4,] 2.977302e-06  1.981321e-05  5.607329e-06  1.692364e-05  9.844239e-06
#>  [5,] 7.645457e-06  4.953990e-06  4.930249e-06  2.809378e-06  5.167194e-06
#>  [6,] 6.033371e-06 -7.399555e-06  6.992680e-07 -7.088606e-06  7.009835e-07
#>  [7,] 7.013963e-07  2.820100e-05  6.125625e-06  2.398200e-05  1.236333e-05
#>  [8,] 2.573625e-06 -7.916154e-06  6.096068e-07 -9.578909e-06 -1.496493e-06
#>  [9,] 5.296881e-06  2.057915e-05  6.848628e-06  1.732574e-05  1.093608e-05
#> [10,] 5.123667e-06 -7.103009e-06  2.801884e-06 -8.137500e-06 -2.409596e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -1.708491e-04 -2.302909e-05 -1.813157e-04 -5.253849e-04 -1.173733e-03
#>  [2,]  2.498488e-04  2.058340e-05  2.262626e-04  5.089352e-04  9.525636e-04
#>  [3,]  1.587109e-05  1.106406e-07  8.241232e-06 -1.500107e-05 -6.855374e-05
#>  [4,]  1.721747e-04  1.366005e-05  1.460413e-04  3.410270e-04  6.308629e-04
#>  [5,]  4.831486e-05  1.167587e-06  3.862493e-05  4.828493e-05  2.868037e-05
#>  [6,] -2.210846e-05 -7.960114e-06 -2.945534e-05 -1.758538e-04 -4.654087e-04
#>  [7,]  2.413327e-04  1.974366e-05  2.065677e-04  5.158832e-04  9.751711e-04
#>  [8,] -1.470010e-05 -8.462683e-06 -2.695995e-05 -1.393886e-04 -3.901269e-04
#>  [9,]  1.739686e-04  1.388589e-05  1.522920e-04  3.450675e-04  6.361989e-04
#> [10,] -5.485064e-05 -6.706193e-06 -6.011101e-05 -1.916019e-04 -4.322786e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,]  0.01341319  0.02066189  0.01502042  0.01593927  0.01574046  0.009395026
#> [2,] -0.01341319 -0.02066189 -0.01502042 -0.01593927 -0.01574046 -0.009395026
#>             [,7]         [,8]        [,9]       [,10]
#> [1,]  0.01860049  0.009332228  0.01756567  0.01774022
#> [2,] -0.01860049 -0.009332228 -0.01756567 -0.01774022
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4889955 0.4813794 0.2714235 0.3497281 0.4596697 0.5875735 0.1248157
#>  [8] 0.6337587 0.1864784 0.3583222
#> 
#> $B[[2]]
#> [1]  0.09602913 -0.05246282
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 4.201544e-04 6.332616e-04 4.222561e-04 2.580035e-04 5.775165e-04
#>  [6] 5.721941e-04 3.911708e-05 4.282475e-04 4.222466e-04 5.370511e-04
#> 
#> $vB[[2]]
#> [1]  0.009068472 -0.009068472
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#>  [2,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#>  [3,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#>  [4,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#>  [5,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#>  [6,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#>  [7,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#>  [8,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#>  [9,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [10,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [11,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [12,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [13,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [14,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [15,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [16,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [17,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [18,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [19,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [20,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [21,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [22,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [23,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [24,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [25,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [26,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [27,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [28,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [29,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [30,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [31,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [32,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [33,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [34,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [35,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [36,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [37,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [38,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [39,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [40,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#>  [2,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#>  [3,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#>  [4,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#>  [5,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#>  [6,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#>  [7,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#>  [8,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#>  [9,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [10,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [11,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [12,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [13,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [14,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [15,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [16,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [17,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [18,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [19,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [20,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [21,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [22,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [23,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [24,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [25,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [26,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [27,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [28,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [29,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [30,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [31,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [32,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [33,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [34,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [35,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [36,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [37,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [38,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [39,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [40,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#>  [2,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#>  [3,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#>  [4,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#>  [5,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#>  [6,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#>  [7,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#>  [8,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#>  [9,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [10,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [11,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [12,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [13,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [14,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [15,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [16,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [17,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [18,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [19,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [20,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [21,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [22,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [23,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [24,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [25,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [26,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [27,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [28,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [29,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [30,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [31,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [32,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [33,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [34,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [35,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [36,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [37,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [38,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [39,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [40,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#>  [2,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#>  [3,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#>  [4,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#>  [5,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#>  [6,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#>  [7,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#>  [8,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#>  [9,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [10,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [11,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [12,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [13,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [14,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [15,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [16,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [17,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [18,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [19,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [20,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [21,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [22,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [23,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [24,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [25,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [26,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [27,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [28,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [29,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [30,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [31,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [32,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [33,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [34,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [35,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [36,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [37,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [38,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [39,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [40,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#>  [2,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#>  [3,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#>  [4,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#>  [5,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#>  [6,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#>  [7,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#>  [8,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#>  [9,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [10,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [11,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [12,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [13,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [14,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [15,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [16,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [17,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [18,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [19,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [20,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [21,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [22,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [23,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [24,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [25,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [26,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [27,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [28,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [29,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [30,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [31,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [32,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [33,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [34,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [35,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [36,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [37,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [38,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [39,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [40,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#>  [2,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#>  [3,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#>  [4,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#>  [5,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#>  [6,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#>  [7,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#>  [8,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#>  [9,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [10,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [11,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [12,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [13,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [14,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [15,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [16,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [17,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [18,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [19,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [20,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [21,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [22,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [23,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [24,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [25,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [26,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [27,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [28,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [29,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [30,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [31,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [32,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [33,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [34,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [35,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [36,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [37,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [38,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [39,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [40,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8079234 0.7031778 0.5517529 0.6480550 0.8254658 0.8921987 0.5897094
#>  [2,] 0.8100191 0.7119681 0.5489165 0.6418460 0.8324576 0.8986347 0.5907696
#>  [3,] 0.6264346 0.5692313 0.3611463 0.3822029 0.5792402 0.8181778 0.3695614
#>  [4,] 0.6678551 0.5538355 0.4440326 0.4956389 0.6595086 0.8091042 0.4405498
#>  [5,] 0.7814897 0.6492421 0.5326812 0.6355816 0.7915721 0.8818372 0.5399226
#>  [6,] 0.7568929 0.6214627 0.4964094 0.5668562 0.7540139 0.8551988 0.5020047
#>  [7,] 0.7369216 0.6320386 0.3979837 0.4947478 0.7281883 0.8512802 0.4711486
#>  [8,] 0.6632964 0.5704488 0.4238894 0.4817096 0.6926559 0.8163568 0.4446057
#>  [9,] 0.7024919 0.5820632 0.4741016 0.5345194 0.7308715 0.8128726 0.4848659
#> [10,] 0.7836892 0.6482813 0.5175050 0.6123190 0.7832689 0.8874751 0.5361530
#> [11,] 0.6890003 0.6137872 0.4406804 0.4526964 0.6936579 0.8439054 0.4354374
#> [12,] 0.7585023 0.6385543 0.4596701 0.5472366 0.7469086 0.8825909 0.4850753
#> [13,] 0.7434894 0.6256230 0.5189278 0.5991722 0.7646285 0.8520382 0.5321981
#> [14,] 0.7180149 0.5775795 0.4427267 0.5171532 0.7143836 0.8551029 0.4377322
#> [15,] 0.7658458 0.6586953 0.5115437 0.6239440 0.7582586 0.8421273 0.5496468
#> [16,] 0.7409496 0.6346186 0.4381865 0.4976782 0.7121993 0.8564356 0.4860781
#> [17,] 0.6826200 0.6057457 0.3857842 0.4290861 0.6864453 0.8230417 0.4586617
#> [18,] 0.7356218 0.6140715 0.5090029 0.5834048 0.7612009 0.8488578 0.5195768
#> [19,] 0.7957050 0.6627590 0.5457887 0.6402834 0.8011660 0.8941716 0.5521137
#> [20,] 0.6985284 0.5363545 0.3829273 0.4496675 0.7208225 0.8215092 0.4218636
#> [21,] 0.7340282 0.6330383 0.3669618 0.4906873 0.7042454 0.8575437 0.4633264
#> [22,] 0.7232828 0.6352551 0.3864740 0.4597795 0.7005391 0.8661238 0.4473500
#> [23,] 0.5667656 0.4249702 0.2908933 0.2891508 0.5744546 0.7272903 0.3203789
#> [24,] 0.7754936 0.6467759 0.5323264 0.6125167 0.7888542 0.8751563 0.5332677
#> [25,] 0.7909368 0.6727759 0.5018994 0.6045370 0.7934136 0.9070895 0.5223236
#> [26,] 0.7254403 0.6136656 0.4745260 0.5572756 0.7551671 0.8347128 0.5043845
#> [27,] 0.6171734 0.5563435 0.3721972 0.3987578 0.6491210 0.7138920 0.4392950
#> [28,] 0.8268562 0.7297421 0.5529894 0.6626232 0.8301610 0.9240858 0.5877433
#> [29,] 0.7269805 0.6450081 0.5082325 0.5447850 0.7881788 0.8421853 0.5186718
#> [30,] 0.6617335 0.5263705 0.3637341 0.4094317 0.6978159 0.7908268 0.4176477
#> [31,] 0.7606362 0.6282809 0.4520511 0.5750872 0.7624243 0.8463441 0.5134459
#> [32,] 0.6202400 0.5440677 0.3924918 0.3874465 0.6097234 0.8150899 0.3843385
#> [33,] 0.7481049 0.6352638 0.4131131 0.5149292 0.7281376 0.8740532 0.4691614
#> [34,] 0.7649878 0.6646362 0.5112197 0.6050509 0.7678377 0.8673490 0.5514575
#> [35,] 0.7333125 0.6199769 0.3726150 0.4813308 0.6986082 0.8639213 0.4474351
#> [36,] 0.6628800 0.5278207 0.3628291 0.4082921 0.6979344 0.7860851 0.4242066
#> [37,] 0.7893344 0.6452557 0.4632147 0.5887201 0.7768435 0.8973608 0.5068237
#> [38,] 0.7501884 0.6183052 0.4632045 0.5495893 0.6952595 0.8511437 0.4894472
#> [39,] 0.6579730 0.6052428 0.4119921 0.4439063 0.7070662 0.7416937 0.4754850
#> [40,] 0.6654214 0.5647632 0.3924133 0.4355032 0.6904744 0.7698183 0.4409156
#>            [,8]      [,9]     [,10]
#>  [1,] 0.9513755 0.6872053 0.6683689
#>  [2,] 0.9556545 0.6926150 0.6647687
#>  [3,] 0.8228223 0.4040551 0.4110264
#>  [4,] 0.8604903 0.5103278 0.5269185
#>  [5,] 0.9440525 0.6450619 0.6448822
#>  [6,] 0.9176767 0.5720847 0.6045279
#>  [7,] 0.9024017 0.5416386 0.4979487
#>  [8,] 0.8782076 0.5147807 0.4919828
#>  [9,] 0.8906519 0.5825374 0.5349185
#> [10,] 0.9431878 0.6258619 0.6344212
#> [11,] 0.8849870 0.5081877 0.4914119
#> [12,] 0.9288896 0.5719186 0.5820802
#> [13,] 0.9176932 0.6273716 0.6104080
#> [14,] 0.9247615 0.5306601 0.5243020
#> [15,] 0.9143047 0.6304625 0.6211868
#> [16,] 0.8984218 0.5345190 0.5442690
#> [17,] 0.8831179 0.5225025 0.4573327
#> [18,] 0.9162575 0.6163704 0.5919727
#> [19,] 0.9488318 0.6528497 0.6581307
#> [20,] 0.9068014 0.5456533 0.4412997
#> [21,] 0.9138068 0.5334484 0.5135668
#> [22,] 0.9025576 0.5200350 0.4861762
#> [23,] 0.8083291 0.3829752 0.2889630
#> [24,] 0.9326224 0.6293897 0.6315562
#> [25,] 0.9476319 0.6259944 0.6174006
#> [26,] 0.9086512 0.5914020 0.5633948
#> [27,] 0.8041044 0.4884272 0.3954629
#> [28,] 0.9589509 0.6762255 0.6824283
#> [29,] 0.9175392 0.6307611 0.5475578
#> [30,] 0.8963616 0.5113094 0.4096583
#> [31,] 0.9227715 0.6034954 0.5723595
#> [32,] 0.8258472 0.4265887 0.4375942
#> [33,] 0.9153906 0.5370111 0.5358236
#> [34,] 0.9217905 0.6192471 0.6039109
#> [35,] 0.9080079 0.5150286 0.4892054
#> [36,] 0.8984344 0.5184314 0.4118523
#> [37,] 0.9453103 0.6065430 0.5909422
#> [38,] 0.9114144 0.5477234 0.5711752
#> [39,] 0.8243835 0.5592319 0.4513324
#> [40,] 0.8795115 0.5326767 0.4407460
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5869218 0.4130782
#>  [2,] 0.5870664 0.4129336
#>  [3,] 0.5895582 0.4104418
#>  [4,] 0.5881488 0.4118512
#>  [5,] 0.5890557 0.4109443
#>  [6,] 0.5904280 0.4095720
#>  [7,] 0.5899144 0.4100856
#>  [8,] 0.5879776 0.4120224
#>  [9,] 0.5867696 0.4132304
#> [10,] 0.5901355 0.4098645
#> [11,] 0.5900608 0.4099392
#> [12,] 0.5918187 0.4081813
#> [13,] 0.5869189 0.4130811
#> [14,] 0.5929340 0.4070660
#> [15,] 0.5854539 0.4145461
#> [16,] 0.5902936 0.4097064
#> [17,] 0.5881242 0.4118758
#> [18,] 0.5874525 0.4125475
#> [19,] 0.5894339 0.4105661
#> [20,] 0.5919993 0.4080007
#> [21,] 0.5909633 0.4090367
#> [22,] 0.5913624 0.4086376
#> [23,] 0.5900750 0.4099250
#> [24,] 0.5890623 0.4109377
#> [25,] 0.5908831 0.4091169
#> [26,] 0.5873651 0.4126349
#> [27,] 0.5812550 0.4187450
#> [28,] 0.5884315 0.4115685
#> [29,] 0.5860136 0.4139864
#> [30,] 0.5904362 0.4095638
#> [31,] 0.5885999 0.4114001
#> [32,] 0.5895179 0.4104821
#> [33,] 0.5917967 0.4082033
#> [34,] 0.5866453 0.4133547
#> [35,] 0.5920582 0.4079418
#> [36,] 0.5900718 0.4099282
#> [37,] 0.5920011 0.4079989
#> [38,] 0.5904235 0.4095765
#> [39,] 0.5806956 0.4193044
#> [40,] 0.5868745 0.4131255
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]       [,2]         [,3]         [,4]      [,5]      [,6]
#>  [1,] 1.4365731  0.8624764  0.207755828  0.610500480 1.5538268 2.1133992
#>  [2,] 1.4501340  0.9049618  0.196293867  0.583385546 1.6031459 2.1821453
#>  [3,] 0.5169513  0.2787156 -0.570392123 -0.480208598 0.3196549 1.5040502
#>  [4,] 0.6984999  0.2161801 -0.224811523 -0.017444962 0.6611052 1.4442001
#>  [5,] 1.2743685  0.6157096  0.130911602  0.556237772 1.3344280 2.0099440
#>  [6,] 1.1357194  0.4957612 -0.014362642  0.269036029 1.1201360 1.7759722
#>  [7,] 1.0300293  0.5409726 -0.413873324 -0.021009698 0.9854501 1.7446775
#>  [8,] 0.6780190  0.2836824 -0.306827008 -0.073194458 0.8125654 1.4918567
#>  [9,] 0.8591924  0.3312487 -0.103686497  0.138297751 0.9990487 1.4687846
#> [10,] 1.2872965  0.6114930  0.070048432  0.457070412 1.2848186 2.0652059
#> [11,] 0.7954497  0.4632598 -0.238401087 -0.189781838 0.8172763 1.6875785
#> [12,] 1.1444856  0.5690949 -0.161670955  0.189511547 1.0821923 2.0171975
#> [13,] 1.0641844  0.5134845  0.075747510  0.402017144 1.1782252 1.7506768
#> [14,] 0.9346359  0.3128448 -0.230103235  0.068639784 0.9167702 1.7751983
#> [15,] 1.1850012  0.6574854  0.046182908  0.506322445 1.1431558 1.6741420
#> [16,] 1.0509098  0.5520826 -0.248525313 -0.009287271 0.9060895 1.7859957
#> [17,] 0.7658386  0.4294640 -0.465068021 -0.285580995 0.7835529 1.5370927
#> [18,] 1.0233353  0.4644594  0.036015434  0.336766391 1.1592745 1.7256708
#> [19,] 1.3596633  0.6756137  0.183669520  0.576594279 1.3935978 2.1340790
#> [20,] 0.8403001  0.1456752 -0.477141701 -0.202014201 0.9485449 1.5266058
#> [21,] 1.0151571  0.5452735 -0.545273202 -0.037255148 0.8675966 1.7950367
#> [22,] 0.9608045  0.5548286 -0.462158210 -0.161230255 0.8498664 1.8671123
#> [23,] 0.2686667 -0.3024026 -0.891049653 -0.899511876 0.3000496 0.9809178
#> [24,] 1.2395954  0.6048970  0.129486227  0.457903617 1.3180325 1.9473402
#> [25,] 1.3305812  0.7207670  0.007597451  0.424405661 1.3456257 2.2786046
#> [26,] 0.9716100  0.4627467 -0.101984363  0.230112291 1.1263633 1.6194033
#> [27,] 0.4775679  0.2263354 -0.522802295 -0.410643780 0.6151776 0.9143624
#> [28,] 1.5635081  0.9933145  0.212756390  0.675005964 1.5867685 2.4992014
#> [29,] 0.9793564  0.5971680  0.032932811  0.179621209 1.3139825 1.6745785
#> [30,] 0.6710287  0.1055800 -0.559193616 -0.366315304 0.8369189 1.3299163
#> [31,] 1.1561705  0.5248491 -0.192386855  0.302637735 1.1660171 1.7062106
#> [32,] 0.4905669  0.1767294 -0.436849882 -0.458058615 0.4461498 1.4834287
#> [33,] 1.0885305  0.5548659 -0.351110925  0.059734597 0.9851942 1.9372815
#> [34,] 1.1802225  0.6840241  0.044886413  0.426555749 1.1961416 1.8777197
#> [35,] 1.0114942  0.4894500 -0.521014785 -0.074711582 0.8406791 1.8482480
#> [36,] 0.6761550  0.1113980 -0.563105837 -0.371030246 0.8374808 1.3014870
#> [37,] 1.3209182  0.5982495 -0.147407578  0.358676677 1.2473654 2.1682380
#> [38,] 1.0996174  0.4823609 -0.147448727  0.199011343 0.8248246 1.7435995
#> [39,] 0.6542741  0.4273588 -0.355736206 -0.225323181 0.8811778 1.0547904
#> [40,] 0.6875486  0.2605165 -0.437179086 -0.259432567 0.8023381 1.2072856
#>               [,7]     [,8]        [,9]        [,10]
#>  [1,]  0.362764194 2.973782  0.78708617  0.700817193
#>  [2,]  0.367147883 3.070385  0.81237342  0.684618601
#>  [3,] -0.534098688 1.535587 -0.38859709 -0.359723870
#>  [4,] -0.238930987 1.819369  0.04131717  0.107778051
#>  [5,]  0.160030952 2.825767  0.59740274  0.596618211
#>  [6,]  0.008019036 2.411191  0.29036182  0.424367575
#>  [7,] -0.115534064 2.224199  0.16694089 -0.008205434
#>  [8,] -0.222490506 1.975565  0.05914009 -0.032071527
#>  [9,] -0.060554925 2.097418  0.33319840  0.139901834
#> [10,]  0.144864749 2.809514  0.51450474  0.551231335
#> [11,] -0.259700293 2.040528  0.03275367 -0.034355974
#> [12,] -0.059716720 2.569756  0.28968315  0.331318538
#> [13,]  0.128971008 2.411409  0.52095756  0.449027472
#> [14,] -0.250370947 2.508873  0.12279455  0.097284786
#> [15,]  0.199243646 2.367366  0.53420128  0.494588621
#> [16,] -0.055701974 2.179810  0.13829587  0.177540734
#> [17,] -0.165731657 2.022293  0.09007084 -0.171085149
#> [18,]  0.078347345 2.392551  0.47417075  0.372126356
#> [19,]  0.209214432 2.920113  0.63158891  0.654974899
#> [20,] -0.315127903 2.275190  0.18312335 -0.235888831
#> [21,] -0.146958472 2.361028  0.13399389  0.054280495
#> [22,] -0.211383724 2.225972  0.08018295 -0.055309378
#> [23,] -0.752031285 1.439190 -0.47693893 -0.900425949
#> [24,]  0.133267777 2.627688  0.52959952  0.538898938
#> [25,]  0.089353777 2.895668  0.51507046  0.478529664
#> [26,]  0.017538383 2.297276  0.36976418  0.254951109
#> [27,] -0.244023807 1.412147 -0.04629953 -0.424405856
#> [28,]  0.354644209 3.151070  0.73647933  0.764953625
#> [29,]  0.074722049 2.409372  0.53548311  0.190808114
#> [30,] -0.332437143 2.157436  0.04524528 -0.365377940
#> [31,]  0.053796765 2.480613  0.42005073  0.291484394
#> [32,] -0.471173552 1.556477 -0.29578289 -0.250931515
#> [33,] -0.123511244 2.381306  0.14831554  0.143540550
#> [34,]  0.206561448 2.466927  0.48635387  0.421787312
#> [35,] -0.211039525 2.289550  0.06013250 -0.043185113
#> [36,] -0.305528260 2.179949  0.07375911 -0.356313409
#> [37,]  0.027296412 2.849839  0.43280391  0.367861804
#> [38,] -0.042217492 2.331029  0.19147661  0.286647702
#> [39,] -0.098138509 1.546333  0.23804555 -0.195288715
#> [40,] -0.237446809 1.987812  0.13089342 -0.238135154
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] 0.09869760 -0.2525571
#>  [2,] 0.09846932 -0.2533818
#>  [3,] 0.12261020 -0.2395291
#>  [4,] 0.11476412 -0.2415539
#>  [5,] 0.10545131 -0.2546118
#>  [6,] 0.11443003 -0.2513051
#>  [7,] 0.11545278 -0.2481587
#>  [8,] 0.11239467 -0.2432163
#>  [9,] 0.10670363 -0.2439231
#> [10,] 0.10918676 -0.2553390
#> [11,] 0.11613950 -0.2480771
#> [12,] 0.11643522 -0.2550536
#> [13,] 0.10308838 -0.2481543
#> [14,] 0.11980030 -0.2563073
#> [15,] 0.10000721 -0.2451960
#> [16,] 0.11695523 -0.2482240
#> [17,] 0.11306226 -0.2431539
#> [18,] 0.10482598 -0.2486179
#> [19,] 0.10567742 -0.2559481
#> [20,] 0.12056608 -0.2516705
#> [21,] 0.11816541 -0.2497834
#> [22,] 0.11892559 -0.2506748
#> [23,] 0.12913158 -0.2351439
#> [24,] 0.10704598 -0.2530444
#> [25,] 0.10995519 -0.2576619
#> [26,] 0.10671934 -0.2463639
#> [27,] 0.10499256 -0.2229350
#> [28,] 0.10130797 -0.2561770
#> [29,] 0.10000771 -0.2475022
#> [30,] 0.11889799 -0.2468707
#> [31,] 0.10919307 -0.2489875
#> [32,] 0.12243191 -0.2395408
#> [33,] 0.11893094 -0.2524667
#> [34,] 0.10182212 -0.2482922
#> [35,] 0.12087212 -0.2516082
#> [36,] 0.11793337 -0.2463288
#> [37,] 0.11450324 -0.2577408
#> [38,] 0.11488669 -0.2508297
#> [39,] 0.09994911 -0.2256806
#> [40,] 0.11065529 -0.2404044
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5869218  0.5869218
#>  [2,] -0.5870664  0.5870664
#>  [3,]  0.4104418 -0.4104418
#>  [4,]  0.4118512 -0.4118512
#>  [5,]  0.4109443 -0.4109443
#>  [6,] -0.5904280  0.5904280
#>  [7,]  0.4100856 -0.4100856
#>  [8,] -0.5879776  0.5879776
#>  [9,]  0.4132304 -0.4132304
#> [10,]  0.4098645 -0.4098645
#> [11,] -0.5900608  0.5900608
#> [12,]  0.4081813 -0.4081813
#> [13,]  0.4130811 -0.4130811
#> [14,]  0.4070660 -0.4070660
#> [15,] -0.5854539  0.5854539
#> [16,] -0.5902936  0.5902936
#> [17,] -0.5881242  0.5881242
#> [18,]  0.4125475 -0.4125475
#> [19,]  0.4105661 -0.4105661
#> [20,]  0.4080007 -0.4080007
#> [21,] -0.5909633  0.5909633
#> [22,]  0.4086376 -0.4086376
#> [23,]  0.4099250 -0.4099250
#> [24,] -0.5890623  0.5890623
#> [25,]  0.4091169 -0.4091169
#> [26,]  0.4126349 -0.4126349
#> [27,] -0.5812550  0.5812550
#> [28,] -0.5884315  0.5884315
#> [29,] -0.5860136  0.5860136
#> [30,]  0.4095638 -0.4095638
#> [31,]  0.4114001 -0.4114001
#> [32,]  0.4104821 -0.4104821
#> [33,]  0.4082033 -0.4082033
#> [34,] -0.5866453  0.5866453
#> [35,]  0.4079418 -0.4079418
#> [36,]  0.4099282 -0.4099282
#> [37,]  0.4079989 -0.4079989
#> [38,] -0.5904235  0.5904235
#> [39,] -0.5806956  0.5806956
#> [40,] -0.5868745  0.5868745
#> 
#> $L
#> [1] 0.7055369 0.6662815 0.7088077 0.7056532 0.6859926 0.6797700
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5507246