Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11         V12          V13          V14
#>  [1,] -0.3413417 -0.195007729 -0.18130202 -0.08036513 -0.212519306 -0.176220248
#>  [2,] -0.3050230 -0.041631245 -0.18649093 -0.13264024 -0.075342594 -0.060679978
#>  [3,] -0.1985643 -0.134732494 -0.12332358 -0.15682374 -0.061184202 -0.052355109
#>  [4,] -0.3062417 -0.193419938 -0.20352891 -0.17654063 -0.131593495 -0.092280080
#>  [5,] -0.4051355 -0.059321314 -0.03920827 -0.03592969 -0.128914002 -0.075526244
#>  [6,] -0.1890351 -0.067353458 -0.07932311 -0.09001382 -0.073122301 -0.099806904
#>  [7,] -0.3457778 -0.122326890 -0.02501219 -0.10130893 -0.078544332 -0.004229963
#>  [8,] -0.2472240 -0.159254386 -0.02775558  0.01651409  0.002797101 -0.125314583
#>  [9,] -0.3259313 -0.004389945 -0.05933054  0.02339680 -0.102331146  0.018854919
#> [10,] -0.3641932 -0.052789069  0.02202589 -0.03373286 -0.023199942  0.008712092
#>               V15         V16         V17         V18         V19         V2
#>  [1,] -0.17574664 -0.19589589 -0.07533004 -0.12985964 -0.12332811 -0.3782274
#>  [2,] -0.00333678 -0.07699147 -0.07993787  0.02400646  0.17084966 -0.2967517
#>  [3,] -0.04847201 -0.13195929 -0.10583602 -0.03529163  0.14746100 -0.2299582
#>  [4,] -0.08916350 -0.04856612 -0.04059787 -0.08801160 -0.09419264 -0.2380731
#>  [5,] -0.01631842  0.05973158  0.05248277  0.17777824  0.22792295 -0.3650207
#>  [6,] -0.09479785  0.01718461 -0.12891624  0.02394177 -0.04173508 -0.2671485
#>  [7,]  0.04723490  0.06483971  0.07952772 -0.02357199  0.15930367 -0.3683243
#>  [8,] -0.07617072 -0.01821762 -0.09873921 -0.11637872 -0.09227132 -0.2026531
#>  [9,] -0.15274875 -0.08605123 -0.09871384 -0.02673274 -0.07165664 -0.2795027
#> [10,] -0.12056677 -0.10661621 -0.06345922 -0.03014819  0.08303625 -0.2691172
#>                V20         V21          V22         V23         V24         V25
#>  [1,] 0.0001797469  0.01083228  0.015281729 -0.01724001  0.02836703 -0.07012514
#>  [2,] 0.1371645577  0.17257666  0.149861021  0.08426144  0.15190644  0.19559493
#>  [3,] 0.0403480644  0.08491266  0.067417178  0.11891182  0.20560572  0.13160930
#>  [4,] 0.0018088551  0.08868989  0.106202081  0.21430936  0.18510076  0.09473367
#>  [5,] 0.1968190644  0.25910134  0.189625064  0.20579020  0.25679826  0.25083641
#>  [6,] 0.1411961680  0.02527849  0.105525768  0.03959968  0.20196195  0.12973833
#>  [7,] 0.2026234698  0.19698907  0.215172900  0.10833597  0.01253876  0.15182936
#>  [8,] 0.0474268061 -0.06607914 -0.021280309  0.01832439 -0.01723172  0.02662371
#>  [9,] 0.0327312808  0.09524421 -0.005488071  0.17665741  0.10081112  0.16968160
#> [10,] 0.0485072654  0.05807318  0.235770024  0.21596112  0.18718989  0.24654806
#>              V26         V27         V28        V29         V3          V30
#>  [1,]  0.1107014  0.10332335  0.11728171 0.18571494 -0.3564250  0.166326427
#>  [2,]  0.1122851  0.07917411  0.13328600 0.14881763 -0.2838548  0.004200034
#>  [3,]  0.1702170  0.06321989  0.02739543 0.02848513 -0.3035947  0.029673958
#>  [4,]  0.2419304  0.11206128  0.31381413 0.26061092 -0.2937775  0.197818119
#>  [5,]  0.2706249  0.31100896  0.26763617 0.13371624 -0.3171540  0.139516225
#>  [6,]  0.1191949  0.12250565  0.04368950 0.04265538 -0.2199384  0.002243275
#>  [7,]  0.1640480  0.17917483  0.08904484 0.08719534 -0.2823461  0.133481973
#>  [8,] -0.0570647 -0.13053095 -0.07588022 0.05065462 -0.2278315  0.037106949
#>  [9,]  0.1912846  0.10377130  0.12659965 0.07585498 -0.3318212 -0.017787927
#> [10,]  0.2855001  0.32704499  0.28457348 0.28093672 -0.2242050  0.152825481
#>               V31          V32         V33          V34          V35
#>  [1,]  0.02020191 -0.103543206  0.02448739  0.119660267  0.050767812
#>  [2,]  0.01190436 -0.020746842  0.08870668  0.074481807 -0.054829433
#>  [3,]  0.08111797 -0.073628026 -0.02181849  0.005666558 -0.078807188
#>  [4,]  0.02305918  0.105858014 -0.03563321  0.051007695 -0.089339551
#>  [5,]  0.01049772  0.005239829  0.02808780 -0.127686670 -0.087477236
#>  [6,] -0.12591994  0.065625819  0.00553509 -0.078691796 -0.038465334
#>  [7,]  0.10375365 -0.003716729  0.03106220  0.002606067  0.098683839
#>  [8,]  0.02815460 -0.071268665 -0.09105337 -0.110741056 -0.004399991
#>  [9,]  0.09677925  0.033799998  0.01178675 -0.015137733 -0.035400190
#> [10,]  0.06046832  0.090917472  0.07283802  0.063153172 -0.018664131
#>               V36         V37         V38          V39         V4         V40
#>  [1,]  0.02575983  0.07735283 -0.08194513 -0.042865551 -0.2110727 -0.02317657
#>  [2,] -0.05831888 -0.13004544 -0.12107303 -0.121028253 -0.3702016 -0.02628908
#>  [3,] -0.10664467 -0.15752335 -0.18286018 -0.122315039 -0.2678744 -0.12266353
#>  [4,] -0.10337832  0.05062973 -0.04026060  0.038710498 -0.2178174 -0.06408979
#>  [5,] -0.05644166 -0.21016614 -0.06382874  0.016710429 -0.3763333 -0.01993997
#>  [6,] -0.03515210 -0.01488556 -0.15059423 -0.154291730 -0.2407552 -0.11987098
#>  [7,]  0.11627956  0.02452947  0.07561107 -0.012265706 -0.2402252  0.01357650
#>  [8,] -0.03879388 -0.05927740 -0.08144700  0.002998155 -0.2833145 -0.06349645
#>  [9,] -0.06487637 -0.12210404 -0.07904114 -0.132137614 -0.2749556 -0.07679723
#> [10,] -0.08434836 -0.12297957 -0.01853761 -0.078804410 -0.2935198 -0.10896970
#>                V41         V42           V43         V44         V45
#>  [1,] -0.130925293 -0.07438367 -1.106448e-01 -0.08821629 -0.01945125
#>  [2,] -0.051118997 -0.10741385 -8.695398e-05 -0.02613149 -0.04219262
#>  [3,] -0.049328939 -0.02715626 -6.510779e-02 -0.16093426 -0.09805652
#>  [4,] -0.063395253 -0.11425052 -7.963045e-02 -0.08230323 -0.06029651
#>  [5,] -0.152266271 -0.11442056 -4.194956e-02 -0.09572166 -0.16592830
#>  [6,] -0.168011499 -0.02399061 -1.907233e-01 -0.03809504 -0.25332197
#>  [7,] -0.063165266 -0.11415627 -5.731255e-02 -0.11492294 -0.20069560
#>  [8,] -0.127956047 -0.09698888 -1.199865e-02 -0.16990077 -0.01441753
#>  [9,] -0.003671049 -0.05689125 -1.185356e-01 -0.19775847 -0.11779012
#> [10,] -0.109937129  0.02000772 -1.372308e-01 -0.10229410 -0.12372388
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.06890557 -0.22786642 -0.1793113 -0.2340051 -0.2487689 -0.2421425
#>  [2,] -0.18686228 -0.11349903 -0.1928577 -0.2830930 -0.2382410 -0.2683019
#>  [3,] -0.07204220 -0.24583338 -0.1509546 -0.3412094 -0.2658208 -0.3675858
#>  [4,] -0.24805751 -0.21336933 -0.1703301 -0.3035378 -0.3007102 -0.3620723
#>  [5,] -0.17624205 -0.24594015 -0.3086291 -0.3078547 -0.2649041 -0.3614015
#>  [6,] -0.18797406 -0.19305744 -0.2448539 -0.2554689 -0.2200658 -0.3658373
#>  [7,] -0.14995679 -0.08213898 -0.3054380 -0.2536354 -0.2662430 -0.3212691
#>  [8,] -0.16154460 -0.24702275 -0.2630217 -0.1895881 -0.2134790 -0.2093086
#>  [9,] -0.08904429 -0.17340722 -0.2935406 -0.3204752 -0.1120186 -0.2687246
#> [10,] -0.20902270 -0.23565707 -0.1207735 -0.2731142 -0.1464608 -0.4092053
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3385620 -0.2911420 -0.3743087 -0.3933083 -0.3581393 -0.3303303
#>  [2,] -0.2995153 -0.2869158 -0.3855265 -0.2485678 -0.2603369 -0.3664621
#>  [3,] -0.3414694 -0.3062727 -0.2415988 -0.2344211 -0.4038369 -0.3404527
#>  [4,] -0.2942403 -0.2482543 -0.2817893 -0.2559916 -0.3269285 -0.3592837
#>  [5,] -0.4721876 -0.3752747 -0.3809242 -0.5393537 -0.3587630 -0.5136299
#>  [6,] -0.3026535 -0.3231221 -0.2588644 -0.2759393 -0.3396648 -0.2885317
#>  [7,] -0.2796439 -0.3921861 -0.4133102 -0.3089525 -0.3719574 -0.4498453
#>  [8,] -0.2876161 -0.3562491 -0.3576283 -0.3686576 -0.3102860 -0.3510752
#>  [9,] -0.3540069 -0.3660354 -0.1982008 -0.2407390 -0.2353090 -0.2317719
#> [10,] -0.2568151 -0.3050967 -0.3112180 -0.3480111 -0.3201645 -0.2564645
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.4094614 -0.3679731 -0.3551894 -0.13213896 -0.2601135 -0.20104110
#>  [2,] -0.3371274 -0.4188271 -0.2710384 -0.30790675 -0.3450947 -0.15460769
#>  [3,] -0.2528427 -0.3124938 -0.2682228 -0.12139841 -0.3503295 -0.11669620
#>  [4,] -0.3951673 -0.3146667 -0.4303699 -0.16750550 -0.3749743 -0.09249869
#>  [5,] -0.4672468 -0.3780885 -0.4684650 -0.28804033 -0.5590264 -0.29614421
#>  [6,] -0.2428478 -0.3171626 -0.3447718 -0.23474234 -0.2727849 -0.19535699
#>  [7,] -0.4674678 -0.3350993 -0.2881822 -0.22707050 -0.3174518 -0.19033455
#>  [8,] -0.3035844 -0.2861483 -0.3778205 -0.09804954 -0.3158145 -0.11350487
#>  [9,] -0.2443226 -0.2483011 -0.3071730 -0.24771745 -0.2816542 -0.13891812
#> [10,] -0.2642272 -0.2803039 -0.2935012 -0.16201706 -0.3792301 -0.26052377
#>                V8           V9
#>  [1,] -0.18062251 -0.121059480
#>  [2,] -0.20042419 -0.156822153
#>  [3,] -0.16792732 -0.075268795
#>  [4,] -0.23180211 -0.200975769
#>  [5,] -0.22871777 -0.144808872
#>  [6,] -0.17852659 -0.093045540
#>  [7,] -0.14734945 -0.197680930
#>  [8,] -0.07815149 -0.178121404
#>  [9,] -0.12582544 -0.008961373
#> [10,] -0.14194693 -0.170007827
#> 
#> $W[[2]]
#>            [,1]       [,2]        [,3]       [,4]       [,5]        [,6]
#> [1,] -0.2451731 -0.1598880 0.002311415 -0.0590752 0.02049883 -0.02497932
#> [2,]  0.1587968  0.1263818 0.089152219  0.1016005 0.06157781  0.09765080
#>            [,7]        [,8]         [,9]       [,10]
#> [1,] -0.0531244 -0.04825786 -0.001549519 -0.04294830
#> [2,]  0.1887582  0.11932949  0.069008223  0.06830545
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10         V11         V12         V13         V14
#>  [1,] 0.0001909426 0.0018446600 0.002242824 0.002194582 0.001672564 0.001029868
#>  [2,] 0.0002996592 0.0022100481 0.002396602 0.002311845 0.002463709 0.002653327
#>  [3,] 0.0002210596 0.0014566006 0.001371057 0.001383196 0.002029842 0.002866697
#>  [4,] 0.0002113851 0.0013916506 0.001399261 0.001454121 0.001966486 0.002596783
#>  [5,] 0.0001549537 0.0008828322 0.000846626 0.001067976 0.001539061 0.001911058
#>  [6,] 0.0002041736 0.0014007000 0.001368498 0.001378622 0.001887831 0.002544702
#>  [7,] 0.0002546365 0.0018165887 0.001903154 0.001885474 0.002174099 0.002495500
#>  [8,] 0.0001231633 0.0009954689 0.001114911 0.001122480 0.001148864 0.001176372
#>  [9,] 0.0002301505 0.0014600461 0.001360412 0.001405333 0.002148799 0.003087915
#> [10,] 0.0001757036 0.0010701051 0.001068441 0.001175945 0.001618606 0.002098833
#>                V15           V16           V17           V18          V19
#>  [1,] 3.881869e-05 -0.0006360073 -0.0006190188 -0.0004747157 0.0001919501
#>  [2,] 2.753905e-03  0.0029767047  0.0032153868  0.0034156762 0.0040983467
#>  [3,] 3.905498e-03  0.0049177380  0.0053214516  0.0055841242 0.0059178310
#>  [4,] 3.441310e-03  0.0042516799  0.0046604168  0.0049147502 0.0052364822
#>  [5,] 2.597034e-03  0.0032188154  0.0035634510  0.0038697724 0.0040813937
#>  [6,] 3.330307e-03  0.0041273505  0.0044786861  0.0047224388 0.0050697709
#>  [7,] 2.836537e-03  0.0032349712  0.0034992908  0.0037245556 0.0042709098
#>  [8,] 1.164572e-03  0.0012501427  0.0013859944  0.0015193490 0.0018308249
#>  [9,] 4.321540e-03  0.0054920788  0.0059655376  0.0062551779 0.0065660485
#> [10,] 2.808219e-03  0.0034538200  0.0038167328  0.0040412638 0.0042798844
#>                 V2          V20         V21           V22          V23
#>  [1,] 0.0002475275 0.0007816659 0.000273869 -0.0007050387 -0.001737277
#>  [2,] 0.0003758629 0.0045396997 0.004653434  0.0047866931  0.004980009
#>  [3,] 0.0002695610 0.0059798412 0.006776583  0.0080882994  0.009505308
#>  [4,] 0.0002658873 0.0051277059 0.005727838  0.0067796495  0.007839948
#>  [5,] 0.0002163139 0.0036281570 0.004225630  0.0054582923  0.006686124
#>  [6,] 0.0002492842 0.0052095044 0.005818698  0.0068277143  0.007946675
#>  [7,] 0.0003230973 0.0045072957 0.004816407  0.0053342086  0.005929235
#>  [8,] 0.0001554029 0.0020265151 0.002085637  0.0021425812  0.002215703
#>  [9,] 0.0002815594 0.0064902504 0.007396819  0.0089299329  0.010544538
#> [10,] 0.0002305701 0.0040080565 0.004563626  0.0055818691  0.006528445
#>                V24          V25          V26          V27          V28
#>  [1,] -0.002160799 -0.002567026 -0.002473729 -0.002130379 -0.001805560
#>  [2,]  0.005051022  0.005154284  0.005334358  0.005215366  0.005235309
#>  [3,]  0.010089726  0.010641804  0.010805778  0.010350870  0.010057506
#>  [4,]  0.008243987  0.008635494  0.008679364  0.008246711  0.007962465
#>  [5,]  0.007229559  0.007637440  0.007562541  0.007124831  0.007144866
#>  [6,]  0.008394181  0.008813012  0.008958848  0.008620741  0.008389829
#>  [7,]  0.006187588  0.006422320  0.006549251  0.006347278  0.006331403
#>  [8,]  0.002241694  0.002246697  0.002308046  0.002299676  0.002300120
#>  [9,]  0.011207934  0.011825994  0.011956405  0.011420177  0.011050479
#> [10,]  0.006885756  0.007198718  0.007167885  0.006765260  0.006586352
#>                V29           V3          V30          V31          V32
#>  [1,] -0.002217653 0.0001928761 -0.002487640 -0.003101845 -0.002251906
#>  [2,]  0.004847102 0.0004116843  0.004465105  0.003702377  0.003283799
#>  [3,]  0.009829808 0.0003828446  0.009367754  0.008790635  0.007244096
#>  [4,]  0.007874718 0.0003648888  0.007524877  0.007131805  0.005865554
#>  [5,]  0.007520962 0.0003106914  0.007355377  0.007104752  0.005785841
#>  [6,]  0.008127506 0.0003406369  0.007683426  0.007133956  0.005924008
#>  [7,]  0.006092649 0.0003835318  0.005717823  0.005096687  0.004330172
#>  [8,]  0.002080900 0.0001731213  0.001831026  0.001473763  0.001310949
#>  [9,]  0.010868205 0.0004154546  0.010385787  0.009859025  0.008088594
#> [10,]  0.006700921 0.0003173252  0.006476077  0.006256669  0.005109081
#>                V33           V34           V35           V36           V37
#>  [1,] -0.002552199 -0.0030721945 -0.0038117774 -0.0040421744 -0.0029894747
#>  [2,]  0.002974253  0.0026099192  0.0024636862  0.0022963798  0.0022580949
#>  [3,]  0.007070823  0.0070218417  0.0074842719  0.0075121875  0.0064130897
#>  [4,]  0.005634484  0.0056179577  0.0060497168  0.0061782977  0.0053461964
#>  [5,]  0.005669080  0.0058376534  0.0064102936  0.0066118501  0.0054434747
#>  [6,]  0.005753996  0.0056470734  0.0059385749  0.0059229483  0.0051054223
#>  [7,]  0.004057094  0.0037951021  0.0038355417  0.0037618583  0.0033730495
#>  [8,]  0.001148851  0.0009528211  0.0008397886  0.0007728341  0.0008275749
#>  [9,]  0.007880989  0.0078746738  0.0084664704  0.0085743552  0.0073205017
#> [10,]  0.004863128  0.0049060855  0.0053937420  0.0056151892  0.0048176072
#>                 V38           V39           V4           V40           V41
#>  [1,] -0.0011912558 -0.0001342194 0.0004276249 -0.0004497827 -0.0004865964
#>  [2,]  0.0023308907  0.0024770993 0.0005219545  0.0021741467  0.0020013733
#>  [3,]  0.0047009094  0.0039055249 0.0003011278  0.0038814359  0.0036258761
#>  [4,]  0.0040154715  0.0035127808 0.0002858923  0.0035521707  0.0032865044
#>  [5,]  0.0040816326  0.0036904666 0.0001669296  0.0037485120  0.0034112796
#>  [6,]  0.0038231605  0.0032334489 0.0002959746  0.0031787680  0.0029777139
#>  [7,]  0.0029161367  0.0028313689 0.0004161458  0.0026473492  0.0024471854
#>  [8,]  0.0009194282  0.0010293574 0.0002230528  0.0009148380  0.0008390724
#>  [9,]  0.0053020783  0.0043829863 0.0002929116  0.0044094677  0.0041142085
#> [10,]  0.0036264434  0.0032518761 0.0002165688  0.0033285288  0.0030540700
#>                V42          V43          V44          V45          V46
#>  [1,] 5.688507e-05 0.0003249135 0.0005287266 0.0014412047 0.0013535839
#>  [2,] 2.130282e-03 0.0021356377 0.0021159947 0.0020877645 0.0017281118
#>  [3,] 3.221075e-03 0.0029214145 0.0026184643 0.0015780238 0.0011535127
#>  [4,] 2.942148e-03 0.0026469152 0.0024460499 0.0015647798 0.0011800068
#>  [5,] 3.031106e-03 0.0026653130 0.0024497110 0.0013087690 0.0009369069
#>  [6,] 2.695568e-03 0.0024668251 0.0022176896 0.0014370870 0.0010767812
#>  [7,] 2.409625e-03 0.0023035994 0.0021995046 0.0018285421 0.0014705609
#>  [8,] 9.086934e-04 0.0009176037 0.0008857765 0.0008724771 0.0007317838
#>  [9,] 3.613736e-03 0.0032456232 0.0029145427 0.0016664081 0.0011970277
#> [10,] 2.710654e-03 0.0024083293 0.0022528953 0.0013948207 0.0010450270
#>                V47          V48          V49           V5          V50
#>  [1,] 0.0007094110 0.0004471390 0.0003669406 0.0003728298 6.501315e-05
#>  [2,] 0.0012843702 0.0009472391 0.0005302303 0.0007199359 1.789311e-04
#>  [3,] 0.0011961364 0.0009550989 0.0004233133 0.0006792187 1.954861e-04
#>  [4,] 0.0012084318 0.0009427734 0.0004399213 0.0006454892 1.814807e-04
#>  [5,] 0.0010980240 0.0008614083 0.0003999283 0.0005225195 1.548874e-04
#>  [6,] 0.0010557710 0.0008335508 0.0003832607 0.0006037827 1.709484e-04
#>  [7,] 0.0012230849 0.0009226293 0.0004807404 0.0006633763 1.750580e-04
#>  [8,] 0.0005554627 0.0004055255 0.0002330274 0.0003120211 7.468028e-05
#>  [9,] 0.0013203960 0.0010582040 0.0004597053 0.0007292484 2.135896e-04
#> [10,] 0.0011106216 0.0008578660 0.0004026216 0.0005482823 1.566443e-04
#>                V51          V52           V53          V54          V55
#>  [1,] 6.935304e-05 8.697707e-05 -1.339229e-06 7.753182e-06 1.654259e-05
#>  [2,] 1.525904e-04 1.361113e-04  9.420568e-05 8.232524e-05 7.520534e-05
#>  [3,] 1.528219e-04 1.085026e-04  1.390180e-04 1.129122e-04 8.883434e-05
#>  [4,] 1.441505e-04 9.922621e-05  1.182753e-04 9.487009e-05 7.758433e-05
#>  [5,] 1.248870e-04 6.743976e-05  1.014198e-04 8.033521e-05 6.361862e-05
#>  [6,] 1.350956e-04 1.006413e-04  1.180430e-04 9.690559e-05 7.705904e-05
#>  [7,] 1.459319e-04 1.191764e-04  1.024741e-04 8.517095e-05 7.369622e-05
#>  [8,] 6.592861e-05 5.867189e-05  3.875331e-05 3.456593e-05 3.017564e-05
#>  [9,] 1.665560e-04 1.120619e-04  1.525673e-04 1.222366e-04 9.525878e-05
#> [10,] 1.254497e-04 8.026505e-05  9.960131e-05 7.805343e-05 6.435413e-05
#>                 V56           V57          V58           V59           V6
#>  [1,] -9.921496e-06 -1.319453e-05 1.938360e-05 -9.242185e-06 0.0001796432
#>  [2,]  6.566441e-05  6.251032e-05 7.466694e-05  6.839713e-05 0.0008265642
#>  [3,]  1.081153e-04  1.067490e-04 8.835770e-05  1.057242e-04 0.0010918891
#>  [4,]  9.134138e-05  9.161619e-05 8.522515e-05  9.267368e-05 0.0009804452
#>  [5,]  8.164531e-05  8.320691e-05 8.391033e-05  9.401658e-05 0.0008410484
#>  [6,]  9.004117e-05  8.857253e-05 7.494065e-05  8.739126e-05 0.0009382858
#>  [7,]  7.400759e-05  7.212562e-05 7.573650e-05  7.716022e-05 0.0008593724
#>  [8,]  2.717237e-05  2.488508e-05 3.044396e-05  2.720963e-05 0.0003708474
#>  [9,]  1.193378e-04  1.184782e-04 9.867450e-05  1.177193e-04 0.0012005767
#> [10,]  7.699832e-05  7.741731e-05 7.936893e-05  8.361814e-05 0.0008374418
#>                 V60           V7           V8           V9
#>  [1,] -8.110200e-06 0.0001947245 0.0005413650 0.0013459395
#>  [2,]  4.979792e-05 0.0010447974 0.0011350317 0.0017609390
#>  [3,]  8.027726e-05 0.0013746820 0.0011727975 0.0012764828
#>  [4,]  6.840411e-05 0.0012203669 0.0010728118 0.0011902532
#>  [5,]  6.306933e-05 0.0010452676 0.0008613084 0.0007762124
#>  [6,]  6.644583e-05 0.0011778225 0.0010323179 0.0012034013
#>  [7,]  5.608794e-05 0.0010800667 0.0010743063 0.0014834828
#>  [8,]  2.010755e-05 0.0004556554 0.0005012093 0.0007829113
#>  [9,]  8.870828e-05 0.0015089616 0.0012589305 0.0013007075
#> [10,]  5.904219e-05 0.0010342714 0.0009037802 0.0009298355
#> 
#> $vW[[2]]
#>            [,1]     [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#> [1,]  0.1279249  0.15834  0.1329151  0.1753597  0.1849859  0.1280553  0.1744401
#> [2,] -0.1279249 -0.15834 -0.1329151 -0.1753597 -0.1849859 -0.1280553 -0.1744401
#>             [,8]      [,9]      [,10]
#> [1,]  0.08492618  0.145796  0.1863917
#> [2,] -0.08492618 -0.145796 -0.1863917
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3327410 0.5034579 0.2446329 0.4705576 0.7767077 0.1989409 0.3901880
#>  [8] 0.2466071 0.3253346 0.6368330
#> 
#> $B[[2]]
#> [1] -0.07834939  0.04568582
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -0.002837542  0.007347741  0.014064347  0.011652298  0.010665417
#>  [6]  0.011666184  0.008887582  0.003120187  0.015636309  0.009895092
#> 
#> $vB[[2]]
#> [1]  0.2072738 -0.2072738
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#>  [2,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#>  [3,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#>  [4,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#>  [5,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#>  [6,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#>  [7,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [8,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#>  [9,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [10,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [11,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [12,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [13,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [14,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [15,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [16,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [17,] 0.0209 0.1853 0.1748 0.1556 0.1476 0.1378 0.2584 0.3827 0.4784 0.5360
#> [18,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [19,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [20,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [21,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [22,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [23,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [24,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [25,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [26,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [27,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [28,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [29,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [30,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [31,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [32,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [33,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [34,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [35,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [36,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [37,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [38,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [39,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [40,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#>  [2,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#>  [3,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#>  [4,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#>  [5,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#>  [6,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#>  [7,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [8,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#>  [9,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [10,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [11,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [12,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [13,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [14,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [15,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [16,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [17,] 0.6192 0.0278 0.7912 0.9264 1.0000 0.9080 0.7435 0.5557 0.3172 0.1295
#> [18,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [19,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [20,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [21,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [22,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [23,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [24,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [25,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [26,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [27,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [28,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [29,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [30,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [31,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [32,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [33,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [34,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [35,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [36,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [37,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [38,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [39,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [40,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#>  [2,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#>  [3,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#>  [4,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#>  [5,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#>  [6,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#>  [7,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [8,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#>  [9,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [10,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [11,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [12,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [13,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [14,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [15,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [16,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [17,] 0.0598 0.2722 0.0115 0.3616 0.3293 0.4855 0.3936 0.1845 0.0342 0.2489
#> [18,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [19,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [20,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [21,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [22,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [23,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [24,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [25,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [26,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [27,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [28,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [29,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [30,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [31,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [32,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [33,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [34,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [35,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [36,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [37,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [38,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [39,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [40,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#>  [2,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#>  [3,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#>  [4,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#>  [5,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#>  [6,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#>  [7,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [8,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#>  [9,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [10,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [11,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [12,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [13,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [14,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [15,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [16,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [17,] 0.3837 0.3514 0.2654 0.0445 0.1760 0.1599 0.0866 0.0590 0.0813 0.0492
#> [18,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [19,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [20,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [21,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [22,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [23,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [24,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [25,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [26,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [27,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [28,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [29,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [30,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [31,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [32,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [33,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [34,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [35,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [36,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [37,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [38,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [39,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [40,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#>  [2,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#>  [3,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#>  [4,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#>  [5,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#>  [6,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#>  [7,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [8,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#>  [9,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [10,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [11,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [12,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [13,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [14,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [15,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [16,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [17,] 0.0417 0.0495 0.0367 0.0115 0.0427 0.0118 0.0133 0.0096 0.0014 0.0049
#> [18,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [19,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [20,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [21,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [22,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [23,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [24,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [25,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [26,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [27,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [28,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [29,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [30,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [31,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [32,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [33,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [34,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [35,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [36,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [37,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [38,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [39,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [40,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#>  [2,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#>  [3,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#>  [4,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#>  [5,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#>  [6,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#>  [7,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [8,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#>  [9,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [10,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [11,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [12,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [13,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [14,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [15,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [16,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [17,] 0.0039 0.0029 0.0078 0.0047 0.0021 0.0766 0.0011 0.1458 0.1430 0.1894
#> [18,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [19,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [20,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [21,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [22,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [23,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [24,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [25,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [26,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [27,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [28,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [29,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [30,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [31,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [32,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [33,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [34,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [35,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [36,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [37,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [38,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [39,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [40,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4629351 0.7762223 0.6386683 0.7690902 0.9250378 0.5831889 0.8327647
#>  [2,] 0.5389219 0.7653468 0.6401625 0.7998134 0.9176549 0.5976151 0.8276445
#>  [3,] 0.4940163 0.6742470 0.5228235 0.7223479 0.8468306 0.5048447 0.7770590
#>  [4,] 0.3687777 0.7306467 0.5720509 0.6761824 0.9087227 0.5231832 0.8180072
#>  [5,] 0.2809101 0.5729727 0.4209331 0.4528616 0.7777357 0.4189846 0.6836253
#>  [6,] 0.5906546 0.8026102 0.7018436 0.8507037 0.9330402 0.6658389 0.8207118
#>  [7,] 0.4643979 0.8209841 0.6838579 0.8113907 0.9538062 0.6488432 0.8715137
#>  [8,] 0.3917417 0.6431390 0.4830168 0.6908201 0.8367201 0.4681066 0.7637776
#>  [9,] 0.4752705 0.7741479 0.6451072 0.7220592 0.9016815 0.6077717 0.8154266
#> [10,] 0.5156485 0.8221901 0.6940864 0.8292580 0.9544063 0.6706957 0.8709026
#> [11,] 0.2913141 0.6568171 0.4905745 0.6003441 0.8690735 0.4512136 0.7569429
#> [12,] 0.5373820 0.8056008 0.6804452 0.8198382 0.9398521 0.6351259 0.8265574
#> [13,] 0.6480875 0.7402485 0.5961032 0.8324510 0.8736977 0.5688048 0.7984073
#> [14,] 0.5616319 0.7701462 0.6447437 0.8230841 0.9165361 0.6078033 0.8158260
#> [15,] 0.3857910 0.5133931 0.3278913 0.5425904 0.6832025 0.3129446 0.6953165
#> [16,] 0.4332870 0.5828474 0.4031327 0.6261645 0.7225096 0.3163551 0.6942152
#> [17,] 0.4508559 0.7358988 0.6006014 0.7093757 0.8801087 0.5774094 0.7926432
#> [18,] 0.3916756 0.6549529 0.4959070 0.6833104 0.8450320 0.4665277 0.7692908
#> [19,] 0.5583805 0.7357342 0.5745711 0.8086261 0.9036125 0.5598761 0.8363707
#> [20,] 0.5203124 0.6920776 0.5471496 0.7443978 0.8455654 0.4987614 0.7287211
#> [21,] 0.5869514 0.7248567 0.5808775 0.7826355 0.8534610 0.5430686 0.7860166
#> [22,] 0.4331299 0.7120230 0.5206619 0.7487401 0.8943983 0.4793359 0.7772204
#> [23,] 0.4348098 0.6395347 0.4867249 0.6849599 0.7796283 0.3852944 0.6963744
#> [24,] 0.4531551 0.7434164 0.5900974 0.7511948 0.9077914 0.5600688 0.7926876
#> [25,] 0.3778433 0.5418802 0.3308346 0.5184881 0.7011097 0.3068015 0.7004275
#> [26,] 0.6388759 0.7886644 0.6414998 0.8339003 0.8945984 0.5815484 0.8205636
#> [27,] 0.4072122 0.5420601 0.3449243 0.5856723 0.7129571 0.3310351 0.7227942
#> [28,] 0.5586558 0.7641291 0.6586577 0.8189773 0.9188876 0.6203844 0.7987902
#> [29,] 0.5785800 0.7501174 0.6187669 0.8157771 0.8834607 0.5816369 0.7958438
#> [30,] 0.5636523 0.7418163 0.6109491 0.7923765 0.8773309 0.5612225 0.7624846
#> [31,] 0.4423573 0.7841272 0.6283446 0.7446397 0.9231675 0.5723366 0.8277709
#> [32,] 0.4024986 0.7596261 0.6124472 0.7418629 0.9287514 0.5659508 0.8248172
#> [33,] 0.6110831 0.7374159 0.5904937 0.8049952 0.8754491 0.5602651 0.8206907
#> [34,] 0.6116549 0.7980137 0.6673170 0.8522425 0.9311316 0.6468854 0.8606117
#> [35,] 0.6301356 0.7972130 0.6752509 0.8516440 0.9260286 0.6488116 0.8493368
#> [36,] 0.5575340 0.6496913 0.4997864 0.6779428 0.7257296 0.4631309 0.7301891
#> [37,] 0.4727156 0.7246783 0.5838359 0.7213075 0.8652912 0.5462137 0.7622476
#> [38,] 0.5292166 0.6848052 0.4953027 0.7102421 0.8319358 0.4966872 0.7687454
#> [39,] 0.4965258 0.7994512 0.6674626 0.8181327 0.9445241 0.6081335 0.8424441
#> [40,] 0.4627555 0.8148236 0.6762372 0.8056502 0.9510673 0.6459528 0.8664396
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3375903 0.6599738 0.8810748
#>  [2,] 0.3840945 0.6817690 0.8954089
#>  [3,] 0.3533148 0.6176201 0.8366640
#>  [4,] 0.3213319 0.6019960 0.8449144
#>  [5,] 0.3290272 0.4606331 0.6497878
#>  [6,] 0.4304820 0.7448720 0.9199996
#>  [7,] 0.3427719 0.7020549 0.9104127
#>  [8,] 0.3047539 0.5751429 0.7927987
#>  [9,] 0.4107197 0.6504675 0.8405899
#> [10,] 0.3749045 0.7217017 0.9195918
#> [11,] 0.2962008 0.5266751 0.7906076
#> [12,] 0.3743172 0.7277855 0.9236391
#> [13,] 0.4159294 0.6887690 0.8953380
#> [14,] 0.3920451 0.7015999 0.9034582
#> [15,] 0.2550869 0.4372612 0.6721277
#> [16,] 0.2836073 0.4517062 0.7179259
#> [17,] 0.4106840 0.6244832 0.8133704
#> [18,] 0.2878095 0.5657162 0.8076642
#> [19,] 0.3435343 0.6743296 0.8970894
#> [20,] 0.3511998 0.6316476 0.8592440
#> [21,] 0.3774904 0.6538528 0.8728574
#> [22,] 0.2971255 0.6138391 0.8807302
#> [23,] 0.2744152 0.5696551 0.8297092
#> [24,] 0.3319138 0.6668846 0.8868770
#> [25,] 0.2525424 0.4345485 0.6648119
#> [26,] 0.3728443 0.7157699 0.9160846
#> [27,] 0.2918526 0.4471388 0.7008891
#> [28,] 0.3998056 0.7220757 0.9108648
#> [29,] 0.3951727 0.6922189 0.8988575
#> [30,] 0.4019263 0.6756439 0.8814354
#> [31,] 0.3458173 0.6320372 0.8673157
#> [32,] 0.3345574 0.6435486 0.8758514
#> [33,] 0.4054122 0.6739675 0.8825702
#> [34,] 0.3938980 0.7370597 0.9254265
#> [35,] 0.3985032 0.7441543 0.9237437
#> [36,] 0.3842432 0.5620812 0.7698120
#> [37,] 0.3610870 0.6363025 0.8359618
#> [38,] 0.3272617 0.5742341 0.8276675
#> [39,] 0.3540885 0.7076553 0.9190778
#> [40,] 0.3446819 0.6983868 0.9072310
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.7948282 0.2051718
#>  [2,] 0.7951772 0.2048228
#>  [3,] 0.7779699 0.2220301
#>  [4,] 0.7858341 0.2141659
#>  [5,] 0.7474575 0.2525425
#>  [6,] 0.8028987 0.1971013
#>  [7,] 0.8039090 0.1960910
#>  [8,] 0.7728264 0.2271736
#>  [9,] 0.7901489 0.2098511
#> [10,] 0.8050944 0.1949056
#> [11,] 0.7720880 0.2279120
#> [12,] 0.8019634 0.1980366
#> [13,] 0.7888443 0.2111557
#> [14,] 0.7965190 0.2034810
#> [15,] 0.7358802 0.2641198
#> [16,] 0.7460380 0.2539620
#> [17,] 0.7843929 0.2156071
#> [18,] 0.7739004 0.2260996
#> [19,] 0.7912626 0.2087374
#> [20,] 0.7793131 0.2206869
#> [21,] 0.7840231 0.2159769
#> [22,] 0.7843992 0.2156008
#> [23,] 0.7665138 0.2334862
#> [24,] 0.7912737 0.2087263
#> [25,] 0.7360916 0.2639084
#> [26,] 0.7943670 0.2056330
#> [27,] 0.7424094 0.2575906
#> [28,] 0.7979864 0.2020136
#> [29,] 0.7917888 0.2082112
#> [30,] 0.7884052 0.2115948
#> [31,] 0.7922793 0.2077207
#> [32,] 0.7934625 0.2065375
#> [33,] 0.7879031 0.2120969
#> [34,] 0.8015634 0.1984366
#> [35,] 0.8010946 0.1989054
#> [36,] 0.7597037 0.2402963
#> [37,] 0.7830040 0.2169960
#> [38,] 0.7725694 0.2274306
#> [39,] 0.8015123 0.1984877
#> [40,] 0.8029825 0.1970175
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]       [,2]          [,3]        [,4]      [,5]         [,6]
#>  [1,] -0.14853192 1.24378603  0.5695886767  1.20318134 2.5128506  0.335877972
#>  [2,]  0.15600310 1.18222018  0.5760694497  1.38512824 2.4109021  0.395537706
#>  [3,] -0.02393599 0.72745715  0.0913573080  0.95613806 1.7099560  0.019379569
#>  [4,] -0.53746422 0.99790635  0.2902238885  0.73628249 2.2981378  0.092799432
#>  [5,] -0.93995152 0.29399026 -0.3189440175 -0.18911540 1.2525193 -0.326943131
#>  [6,]  0.36667208 1.40268890  0.8560921862  1.74013063 2.6343554  0.689424379
#>  [7,] -0.14264986 1.52302914  0.7715581338  1.45907233 3.0276149  0.613958026
#>  [8,] -0.43999682 0.58901473 -0.0679588911  0.80395635 1.6340235 -0.127747056
#>  [9,] -0.09899874 1.23188281  0.5976006102  0.95469884 2.2160495  0.437955391
#> [10,]  0.06261432 1.53125636  0.8192936517  1.58037766 3.0413202  0.711333066
#> [11,] -0.88901030 0.64914224 -0.0377063832  0.40689918 1.8927914 -0.195768267
#> [12,]  0.14980749 1.42167425  0.7558186617  1.51525145 2.7489166  0.554270850
#> [13,]  0.61064297 1.04726074  0.3892542487  1.60309848 1.9340560  0.276976393
#> [14,]  0.24778787 1.20913673  0.5960132585  1.53738409 2.3961876  0.438087726
#> [15,] -0.46503929 0.05358513 -0.7177381216  0.17077544 0.7685288 -0.786388635
#> [16,] -0.26845288 0.33447331 -0.3924289966  0.51579722 0.9569446 -0.770573294
#> [17,] -0.19721305 1.02476006  0.4079717558  0.89235415 1.9934598  0.312147800
#> [18,] -0.44027445 0.64088219 -0.0163725506  0.76902718 1.6961559 -0.134089785
#> [19,]  0.23459181 1.02391371  0.3005259034  1.44110761 2.2380239  0.240659366
#> [20,]  0.08129430 0.80985048  0.1891603025  1.06895325 1.7002345 -0.004954335
#> [21,]  0.35137687 0.96868211  0.3263767780  1.28109124 1.7620086  0.172702433
#> [22,] -0.26909266 0.90522984  0.0826946538  1.09190384 2.1364772 -0.082703567
#> [23,] -0.26225345 0.57334517 -0.0531127085  0.77666016 1.2635015 -0.467135833
#> [24,] -0.18793087 1.06380187  0.3643678918  1.10499449 2.2869611  0.241441228
#> [25,] -0.49871254 0.16791437 -0.7044125298  0.07398616 0.8525876 -0.815115258
#> [26,]  0.57048864 1.31689378  0.5818798554  1.61352556 2.1385968  0.329133171
#> [27,] -0.37550214 0.16863889 -0.6414258498  0.34610314 0.9097897 -0.703507136
#> [28,]  0.23570834 1.17545205  0.6573182436  1.50943388 2.4273282  0.491179968
#> [29,]  0.31694669 1.09923876  0.4843175962  1.48799452 2.0256185  0.329496667
#> [30,]  0.25599814 1.05543034  0.4513036744  1.33931036 1.9673937  0.246125132
#> [31,] -0.23160041 1.28988179  0.5251218491  1.07022522 2.4861829  0.291390884
#> [32,] -0.39506483 1.15063057  0.4576105141  1.05567382 2.5676659  0.265349204
#> [33,]  0.45186740 1.03258055  0.3660066947  1.41781220 1.9500229  0.242238186
#> [34,]  0.45427393 1.37392593  0.6960752425  1.75229896 2.6042038  0.605376538
#> [35,]  0.53279841 1.36896554  0.7320314607  1.74755420 2.5272262  0.613819730
#> [36,]  0.23115976 0.61768253 -0.0008543029  0.74433380 0.9730630 -0.147744630
#> [37,] -0.10924623 0.96778735  0.3385400923  0.95095650 1.8599507  0.185383878
#> [38,]  0.11699971 0.77594336 -0.0187895865  0.89655994 1.5994091 -0.013251377
#> [39,] -0.01389686 1.38286794  0.6967308287  1.50374698 2.8347330  0.439473080
#> [40,] -0.14925461 1.48166285  0.7365329375  1.42198993 2.9671379  0.601296089
#>            [,7]       [,8]       [,9]     [,10]
#>  [1,] 1.6053492 -0.6740512  0.6631775 2.0026482
#>  [2,] 1.5690242 -0.4722047  0.7619135 2.1472215
#>  [3,] 1.2486090 -0.6045002  0.4794591 1.6336130
#>  [4,] 1.5029041 -0.7476574  0.4137889 1.6952585
#>  [5,] 0.7704828 -0.7125882 -0.1577941 0.6181067
#>  [6,] 1.5211776 -0.2798847  1.0714468 2.4423415
#>  [7,] 1.9144093 -0.6509659  0.8571025 2.3186844
#>  [8,] 1.1735026 -0.8247615  0.3028656 1.3418785
#>  [9,] 1.4856645 -0.3609910  0.6210948 1.6626236
#> [10,] 1.9089630 -0.5112332  0.9529184 2.4368144
#> [11,] 1.1359915 -0.8654557  0.1068017 1.3285916
#> [12,] 1.5614228 -0.5137401  0.9834160 2.4928504
#> [13,] 1.3763697 -0.3395063  0.7943704 2.1464653
#> [14,] 1.4883201 -0.4387238  0.8549280 2.2362543
#> [15,] 0.8250935 -1.0716631 -0.2522847 0.7178243
#> [16,] 0.8199004 -0.9266382 -0.1937792 0.9341966
#> [17,] 1.3409323 -0.3611383  0.5086210 1.4720608
#> [18,] 1.2043112 -0.9060466  0.2643944 1.4349037
#> [19,] 1.6314687 -0.6475835  0.7278332 2.1652946
#> [20,] 0.9881438 -0.6137695  0.5392914 1.8090247
#> [21,] 1.3010798 -0.5002137  0.6360182 1.9264632
#> [22,] 1.2495411 -0.8610237  0.4634786 1.9993628
#> [23,] 0.8300920 -0.9723358  0.2804440 1.5835680
#> [24,] 1.3412026 -0.6995419  0.6941279 2.0592306
#> [25,] 0.8493346 -1.0850984 -0.2633168 0.6848122
#> [26,] 1.5201704 -0.5200338  0.9235749 2.3902998
#> [27,] 0.9583645 -0.8864034 -0.2122379 0.8515354
#> [28,] 1.3787502 -0.4062751  0.9547810 2.3242406
#> [29,] 1.3605176 -0.4256203  0.8105132 2.1845936
#> [30,] 1.1663500 -0.3974451  0.7338243 2.0060938
#> [31,] 1.5699108 -0.6374761  0.5409663 1.8774306
#> [32,] 1.5493321 -0.6876440  0.5907995 1.9537176
#> [33,] 1.5210343 -0.3829639  0.7261849 2.0169978
#> [34,] 1.8203801 -0.4309562  1.0307420 2.5184688
#> [35,] 1.7294093 -0.4117058  1.0676736 2.4943353
#> [36,] 0.9955822 -0.4715764  0.2496126 1.2072498
#> [37,] 1.1650419 -0.5706494  0.5593517 1.6284838
#> [38,] 1.2012407 -0.7205962  0.2991476 1.5691855
#> [39,] 1.6765271 -0.6011155  0.8840235 2.4298825
#> [40,] 1.8698388 -0.6424987  0.8396279 2.2802844
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.8344261 -0.5198521
#>  [2,] 0.8355623 -0.5208573
#>  [3,] 0.7670321 -0.4868429
#>  [4,] 0.8002501 -0.4997449
#>  [5,] 0.6623352 -0.4227629
#>  [6,] 0.8673811 -0.5371298
#>  [7,] 0.8730631 -0.5378442
#>  [8,] 0.7483470 -0.4759931
#>  [9,] 0.8195189 -0.5063045
#> [10,] 0.8781436 -0.5403003
#> [11,] 0.7470603 -0.4730786
#> [12,] 0.8620808 -0.5365301
#> [13,] 0.8072567 -0.5107162
#> [14,] 0.8399109 -0.5247672
#> [15,] 0.6161676 -0.4084971
#> [16,] 0.6483781 -0.4292140
#> [17,] 0.7969670 -0.4944851
#> [18,] 0.7515716 -0.4788962
#> [19,] 0.8174456 -0.5151071
#> [20,] 0.7682085 -0.4934591
#> [21,] 0.7886120 -0.5006552
#> [22,] 0.7877690 -0.5037206
#> [23,] 0.7180689 -0.4706609
#> [24,] 0.8178065 -0.5148136
#> [25,] 0.6167832 -0.4089691
#> [26,] 0.8281563 -0.5232965
#> [27,] 0.6391356 -0.4193935
#> [28,] 0.8456870 -0.5280694
#> [29,] 0.8197761 -0.5159657
#> [30,] 0.8057078 -0.5096312
#> [31,] 0.8247768 -0.5139425
#> [32,] 0.8296183 -0.5163059
#> [33,] 0.8054794 -0.5068524
#> [34,] 0.8612407 -0.5348536
#> [35,] 0.8588114 -0.5343384
#> [36,] 0.7000071 -0.4510486
#> [37,] 0.7867208 -0.4965378
#> [38,] 0.7443451 -0.4785316
#> [39,] 0.8607320 -0.5350409
#> [40,] 0.8691754 -0.5358647
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.2051718 -0.2051718
#>  [2,] -0.7951772  0.7951772
#>  [3,] -0.7779699  0.7779699
#>  [4,] -0.7858341  0.7858341
#>  [5,]  0.2525425 -0.2525425
#>  [6,] -0.8028987  0.8028987
#>  [7,] -0.8039090  0.8039090
#>  [8,] -0.7728264  0.7728264
#>  [9,] -0.7901489  0.7901489
#> [10,] -0.8050944  0.8050944
#> [11,]  0.2279120 -0.2279120
#> [12,]  0.1980366 -0.1980366
#> [13,] -0.7888443  0.7888443
#> [14,] -0.7965190  0.7965190
#> [15,]  0.2641198 -0.2641198
#> [16,] -0.7460380  0.7460380
#> [17,]  0.2156071 -0.2156071
#> [18,] -0.7739004  0.7739004
#> [19,] -0.7912626  0.7912626
#> [20,]  0.2206869 -0.2206869
#> [21,] -0.7840231  0.7840231
#> [22,]  0.2156008 -0.2156008
#> [23,]  0.2334862 -0.2334862
#> [24,]  0.2087263 -0.2087263
#> [25,]  0.2639084 -0.2639084
#> [26,] -0.7943670  0.7943670
#> [27,] -0.7424094  0.7424094
#> [28,]  0.2020136 -0.2020136
#> [29,] -0.7917888  0.7917888
#> [30,] -0.7884052  0.7884052
#> [31,]  0.2077207 -0.2077207
#> [32,]  0.2065375 -0.2065375
#> [33,] -0.7879031  0.7879031
#> [34,] -0.8015634  0.8015634
#> [35,] -0.8010946  0.8010946
#> [36,] -0.7597037  0.7597037
#> [37,]  0.2169960 -0.2169960
#> [38,] -0.7725694  0.7725694
#> [39,]  0.1984877 -0.1984877
#> [40,] -0.8029825  0.8029825
#> 
#> $L
#> [1] 0.6978203 0.7205212 0.7575592 0.6895384 0.8352081 1.0261446
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319