Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13         V14
#>  [1,] -0.2479964 -0.04998562 -0.12229194 -0.150374680 -0.17400636 -0.08502420
#>  [2,] -0.2354392 -0.14256146 -0.08264046  0.006802016 -0.02362870  0.07718596
#>  [3,] -0.3066437 -0.17169422 -0.07963572 -0.142890224 -0.08918708 -0.11471758
#>  [4,] -0.1680274 -0.04594081 -0.11124389 -0.204844483 -0.15657151 -0.11144964
#>  [5,] -0.3345654 -0.07917160 -0.02798192 -0.173671181 -0.02707599 -0.17882317
#>  [6,] -0.3178744 -0.08999925 -0.01621452 -0.070485907 -0.02609806 -0.09985477
#>  [7,] -0.2561710 -0.03183095 -0.11774821 -0.072888662 -0.14648790 -0.09820583
#>  [8,] -0.4521620 -0.02493731 -0.20571489 -0.088202015 -0.02275233 -0.13817370
#>  [9,] -0.3929308 -0.17771479 -0.09680879  0.008596649 -0.09869022 -0.07036219
#> [10,] -0.2007312 -0.17951056 -0.06549155 -0.104484823 -0.12917659  0.02242666
#>                V15          V16          V17         V18          V19
#>  [1,] -0.115368376  0.015489022 -0.005132957  0.05620217 -0.002194803
#>  [2,] -0.008579797  0.001703376 -0.060609908 -0.07965971  0.164670199
#>  [3,] -0.006479528 -0.057250660 -0.062196799 -0.05982977  0.006106966
#>  [4,] -0.154765276 -0.026383247 -0.137264351 -0.17402367 -0.140236922
#>  [5,] -0.023070483 -0.137943037  0.037488331  0.07309313 -0.057801450
#>  [6,] -0.153211632  0.050636139 -0.070057481  0.02614502 -0.027186394
#>  [7,] -0.002708898 -0.056319558 -0.117277441 -0.02450842  0.005170600
#>  [8,] -0.138462758  0.037051062 -0.081152556 -0.08989218 -0.021725085
#>  [9,]  0.061910352  0.138401389  0.091638133  0.08223821  0.092942893
#> [10,] -0.028170617 -0.084313485  0.026060655  0.02123744  0.020480797
#>               V2         V20         V21          V22         V23        V24
#>  [1,] -0.2552400  0.04040561 0.034624280  0.004132365  0.14513342 0.07691880
#>  [2,] -0.1770006  0.15105071 0.223079154  0.217960090  0.26816917 0.21053293
#>  [3,] -0.3586941  0.07951401 0.240165004  0.288570430  0.21853249 0.30525906
#>  [4,] -0.2998529 -0.10438895 0.005942837 -0.134183218 -0.16210516 0.01406833
#>  [5,] -0.1442528 -0.04882595 0.003345589  0.131025815  0.09973651 0.05093036
#>  [6,] -0.2696438  0.02294891 0.129517502  0.060326721  0.21245843 0.14507834
#>  [7,] -0.2477273  0.08247385 0.108279543  0.017875859 -0.03699807 0.08609176
#>  [8,] -0.3818197  0.09712877 0.014157080  0.068653155  0.16702067 0.17525796
#>  [9,] -0.3261427  0.21463125 0.111876967  0.276535157  0.29108422 0.26902576
#> [10,] -0.3532618 -0.01970097 0.107331453  0.149237055  0.18932065 0.11081987
#>               V25          V26         V27         V28         V29         V3
#>  [1,]  0.02661717  0.009677201  0.17252169  0.00270075 -0.06260796 -0.2595473
#>  [2,]  0.23572421  0.226568200  0.25995947  0.23890069  0.22483998 -0.2809274
#>  [3,]  0.14952751  0.220307722  0.21095585  0.24404271  0.20689696 -0.2036820
#>  [4,] -0.11062529 -0.091364473  0.06953036  0.02760827 -0.04095059 -0.2528780
#>  [5,]  0.06051840  0.088470340  0.07455869  0.14064376 -0.01372993 -0.2305708
#>  [6,]  0.22427173  0.146278377  0.14824170  0.18968525  0.09540861 -0.2802800
#>  [7,]  0.08071855  0.069986351 -0.03281651 -0.01454869  0.03071638 -0.3326041
#>  [8,]  0.06613285  0.222276283  0.11542835  0.07573820  0.14426725 -0.2614635
#>  [9,]  0.27163506  0.123054002  0.25923171  0.15235339  0.04571111 -0.3455155
#> [10,]  0.25483089  0.296154055  0.20957543  0.12123539  0.11758424 -0.3435849
#>               V30         V31          V32          V33           V34
#>  [1,] -0.09655994 -0.05180394 -0.052243363  0.008419286 -0.0625561307
#>  [2,]  0.23429625  0.04240743  0.079601504  0.062356145  0.1329713445
#>  [3,]  0.04587220 -0.09463649 -0.080281386 -0.005324287 -0.1091495923
#>  [4,] -0.15187531 -0.04065810 -0.127923834 -0.217366480 -0.0795932830
#>  [5,]  0.02767789 -0.03006409  0.050623747 -0.089494527 -0.0955259880
#>  [6,]  0.12888331  0.05449717  0.037524910  0.027786726 -0.0148177619
#>  [7,] -0.05763735  0.01714347 -0.102208414 -0.011106921  0.0002095378
#>  [8,]  0.10123379  0.06550688 -0.091768412 -0.015237478 -0.0114195289
#>  [9,] -0.02919525  0.01479435 -0.040109047 -0.068624907 -0.1008487744
#> [10,]  0.07254252  0.08155855  0.001832663 -0.089681682  0.0268953216
#>                 V35          V36         V37          V38         V39
#>  [1,]  0.0252555099 -0.096302713 -0.01291139  0.019403395  0.04644724
#>  [2,]  0.1344890497  0.182070203  0.03509584  0.178040474  0.06979122
#>  [3,] -0.1204647772  0.013237932  0.04476168 -0.007513957 -0.05321532
#>  [4,] -0.1472443420 -0.155740173 -0.12728876 -0.035178413 -0.13778265
#>  [5,] -0.1040047121  0.010024271 -0.01896925 -0.042668033 -0.11309537
#>  [6,]  0.1013234592  0.091474472  0.03538964  0.068835666  0.11357640
#>  [7,] -0.1004838531 -0.138412167  0.04944062 -0.118387712 -0.08196729
#>  [8,] -0.1867298300 -0.207673936 -0.10240733 -0.048221288 -0.03222675
#>  [9,] -0.0008396255 -0.008175148  0.06431287 -0.141209688 -0.16657239
#> [10,] -0.0428720994 -0.044630223 -0.05313751 -0.087166214 -0.10281794
#>               V4          V40         V41          V42         V43         V44
#>  [1,] -0.2194280  0.010145533 -0.13934823 -0.013036626 -0.15128422 -0.15158964
#>  [2,] -0.2852836  0.003927812 -0.08037059 -0.100374200  0.01519983 -0.13237609
#>  [3,] -0.2883147 -0.100931697 -0.19105277 -0.204441696 -0.12247754 -0.10636540
#>  [4,] -0.1847962 -0.052435552 -0.18741874 -0.183959626 -0.13781773 -0.19752157
#>  [5,] -0.2979120 -0.042599262 -0.18029775 -0.053409137 -0.02165727 -0.19643719
#>  [6,] -0.2137376  0.040558792 -0.07900138 -0.080786873 -0.09920669 -0.06673404
#>  [7,] -0.2875656 -0.154049379 -0.06948413 -0.172219994 -0.17203435 -0.09302647
#>  [8,] -0.3502545 -0.180241909 -0.17661304 -0.099046972 -0.19976185 -0.11844014
#>  [9,] -0.2915768 -0.062810198 -0.05029315 -0.030135985 -0.04537651 -0.06545843
#> [10,] -0.2056450 -0.088646654 -0.01495673  0.006318688 -0.03428235 -0.03857286
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.10945115 -0.10996481 -0.1700597 -0.1402705 -0.2710725 -0.2952537
#>  [2,] -0.10707404 -0.11386647 -0.1187015 -0.2162459 -0.1958929 -0.2259394
#>  [3,] -0.02310726 -0.13525649 -0.1838408 -0.2058080 -0.3377091 -0.2363162
#>  [4,] -0.20058457 -0.11804554 -0.3021456 -0.2048978 -0.2424030 -0.1411237
#>  [5,] -0.20015519 -0.10023969 -0.1234162 -0.1221890 -0.2009828 -0.2357271
#>  [6,] -0.10495280 -0.05498709 -0.2303196 -0.2574730 -0.3373323 -0.1461216
#>  [7,] -0.18968976 -0.14128333 -0.1490455 -0.1968338 -0.3078227 -0.1715263
#>  [8,] -0.13495071 -0.19165984 -0.1966835 -0.3477210 -0.2404532 -0.3049292
#>  [9,] -0.15689326 -0.16614201 -0.2375101 -0.2917239 -0.2161135 -0.2692276
#> [10,] -0.01431739 -0.17895735 -0.2124920 -0.1382172 -0.2855779 -0.2054069
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3179353 -0.2650433 -0.2687905 -0.2374717 -0.3359262 -0.2990211
#>  [2,] -0.2176428 -0.2697992 -0.3866678 -0.3352290 -0.3371631 -0.3794250
#>  [3,] -0.3462209 -0.2568923 -0.2541149 -0.3203397 -0.4270585 -0.2755444
#>  [4,] -0.1795425 -0.3094418 -0.1824583 -0.2385706 -0.1557359 -0.3387271
#>  [5,] -0.3027530 -0.2653667 -0.2185865 -0.2285505 -0.2206286 -0.2920968
#>  [6,] -0.2490057 -0.4021524 -0.3710099 -0.3612368 -0.3285970 -0.3678379
#>  [7,] -0.4031145 -0.3945869 -0.4068048 -0.3859981 -0.3218624 -0.3979301
#>  [8,] -0.3290921 -0.4489598 -0.3061642 -0.3159683 -0.3995502 -0.4605472
#>  [9,] -0.4018232 -0.3006981 -0.3833198 -0.3674019 -0.3073922 -0.3099666
#> [10,] -0.3584471 -0.3419751 -0.3711150 -0.4382448 -0.2547718 -0.2491781
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.2676754 -0.2645900 -0.2048228 -0.2374550 -0.2468487 -0.2723747
#>  [2,] -0.4033528 -0.3140603 -0.3184921 -0.3322514 -0.2078032 -0.3583309
#>  [3,] -0.3244269 -0.3801099 -0.2863708 -0.3229874 -0.1235104 -0.3226466
#>  [4,] -0.2580181 -0.1454831 -0.1842928 -0.2924050 -0.1633362 -0.2806909
#>  [5,] -0.2865913 -0.3047453 -0.3648522 -0.2449737 -0.1530549 -0.2545596
#>  [6,] -0.3640624 -0.4098010 -0.2800099 -0.3917203 -0.1885322 -0.3835868
#>  [7,] -0.3005847 -0.3382462 -0.3062813 -0.3756836 -0.1250947 -0.3439290
#>  [8,] -0.3778532 -0.4171902 -0.4543931 -0.4024119 -0.2788409 -0.3747898
#>  [9,] -0.4547960 -0.3542774 -0.4088106 -0.4016712 -0.2184462 -0.3103018
#> [10,] -0.2937764 -0.3676634 -0.2546150 -0.3172954 -0.2100898 -0.3502727
#>               V7          V8          V9
#>  [1,] -0.2156788 -0.13469965 -0.18279781
#>  [2,] -0.2226375 -0.05998713 -0.09602037
#>  [3,] -0.1522628 -0.10738504 -0.09216498
#>  [4,] -0.2060408 -0.22853627 -0.18366572
#>  [5,] -0.0959156 -0.10035012 -0.17528279
#>  [6,] -0.1361295 -0.18440969 -0.13909279
#>  [7,] -0.1159192 -0.17417010 -0.04469603
#>  [8,] -0.2031409 -0.12340763 -0.14479114
#>  [9,] -0.1257290 -0.21559944 -0.20877149
#> [10,] -0.1655358 -0.15067624 -0.13823796
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]         [,4]        [,5]        [,6]
#> [1,] 0.03749443  0.08146098  0.14643339 -0.085769341 -0.08219215  0.03026482
#> [2,] 0.02398288 -0.07450673 -0.06666945  0.003286869 -0.00483785 -0.03351096
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.03393548 0.001091722 -0.03973396 0.032819837
#> [2,] -0.04808278 0.058554302 -0.03881209 0.009435416
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  9.592964e-05  0.0007160738  0.0007156925  0.0006021466  0.0007162701
#>  [2,]  3.083072e-05  0.0002438206  0.0002181624  0.0001778096  0.0002654056
#>  [3,] -5.472306e-05 -0.0004024532 -0.0005626875 -0.0004365992 -0.0001228408
#>  [4,]  5.412860e-05  0.0005384575  0.0006201974  0.0005239634  0.0004148523
#>  [5,]  2.478180e-04  0.0021017436  0.0023474349  0.0020319191  0.0018603272
#>  [6,]  1.311505e-04  0.0009604134  0.0009904088  0.0008207859  0.0009006673
#>  [7,]  1.159440e-04  0.0008860733  0.0009015296  0.0007577354  0.0008608495
#>  [8,]  2.589586e-04  0.0020912251  0.0023083635  0.0019761624  0.0018604364
#>  [9,]  2.086159e-04  0.0015200878  0.0016072366  0.0012967270  0.0012732973
#> [10,]  1.890106e-04  0.0013993502  0.0014789697  0.0012185649  0.0012334439
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0008802498 0.0011473302 0.0013920274 0.0015351126 0.0017068068
#>  [2,] 0.0003937781 0.0005269287 0.0006286184 0.0006889371 0.0007368473
#>  [3,] 0.0004475584 0.0008003276 0.0009120933 0.0011770583 0.0011357006
#>  [4,] 0.0002402152 0.0002528001 0.0003270595 0.0003695432 0.0005156742
#>  [5,] 0.0014463465 0.0015328614 0.0019158408 0.0020416019 0.0025748943
#>  [6,] 0.0009686561 0.0012081728 0.0014661717 0.0015293399 0.0017150283
#>  [7,] 0.0009948021 0.0012704367 0.0015425197 0.0016836850 0.0018902778
#>  [8,] 0.0015229403 0.0016493409 0.0020522013 0.0021393698 0.0026266720
#>  [9,] 0.0011114100 0.0012863610 0.0015889853 0.0015038360 0.0017225609
#> [10,] 0.0011689204 0.0013949904 0.0017081379 0.0017185166 0.0019766005
#>                V19            V2           V20           V21           V22
#>  [1,] 0.0016820458  1.240363e-04  1.945842e-03  0.0021963013  2.278581e-03
#>  [2,] 0.0006345221  4.250200e-05  7.009659e-04  0.0007663434  7.937042e-04
#>  [3,] 0.0003742383 -5.064681e-05 -5.324601e-05 -0.0001776458 -3.681371e-06
#>  [4,] 0.0007179761  7.381696e-05  9.221080e-04  0.0010968644  1.117004e-03
#>  [5,] 0.0033813894  3.168694e-04  4.288593e-03  0.0049252641  4.976356e-03
#>  [6,] 0.0018565233  1.643729e-04  2.285014e-03  0.0025784239  2.567641e-03
#>  [7,] 0.0019310503  1.484490e-04  2.271584e-03  0.0025668500  2.638010e-03
#>  [8,] 0.0033894959  3.228210e-04  4.331380e-03  0.0049519548  4.962064e-03
#>  [9,] 0.0021312582  2.538373e-04  2.893838e-03  0.0032813599  3.155750e-03
#> [10,] 0.0023335728  2.311101e-04  2.999613e-03  0.0033910460  3.317590e-03
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0023418810 0.0024539883 0.0024251172 0.0024358358 0.0024814636
#>  [2,] 0.0008379545 0.0008997775 0.0009059221 0.0009289724 0.0009202274
#>  [3,] 0.0002234204 0.0004311781 0.0007002684 0.0007601339 0.0005147624
#>  [4,] 0.0010622459 0.0010366512 0.0008723463 0.0008159588 0.0009152383
#>  [5,] 0.0048761559 0.0048508573 0.0043697910 0.0042677402 0.0046612676
#>  [6,] 0.0025507876 0.0026253160 0.0025098824 0.0025088047 0.0026159147
#>  [7,] 0.0026820880 0.0027909963 0.0027261256 0.0027296818 0.0028009119
#>  [8,] 0.0048512789 0.0048283974 0.0043858056 0.0043124370 0.0046942402
#>  [9,] 0.0030490297 0.0031132098 0.0029573346 0.0030197819 0.0032262522
#> [10,] 0.0032330834 0.0032705962 0.0030709555 0.0030809691 0.0032819224
#>                V28          V29           V3          V30          V31
#>  [1,] 0.0024579383 0.0021762690 1.422529e-04 0.0020766409 0.0017636020
#>  [2,] 0.0009106284 0.0008254363 5.741883e-05 0.0007850560 0.0007234843
#>  [3,] 0.0002929854 0.0003464308 2.905888e-06 0.0002665517 0.0006490200
#>  [4,] 0.0010024531 0.0008820543 6.660930e-05 0.0008194726 0.0005057926
#>  [5,] 0.0048921409 0.0042871990 3.142678e-04 0.0040508774 0.0029188208
#>  [6,] 0.0026985706 0.0023996819 1.764488e-04 0.0023099723 0.0018695823
#>  [7,] 0.0028004162 0.0024756948 1.656121e-04 0.0023635621 0.0019625153
#>  [8,] 0.0049238385 0.0043157887 3.244852e-04 0.0041421852 0.0030471730
#>  [9,] 0.0034510755 0.0030463820 2.290094e-04 0.0030706410 0.0023982397
#> [10,] 0.0034452014 0.0030470616 2.317059e-04 0.0029930574 0.0023203268
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0013319002 0.0012692496 0.0012154840 0.0013418357 0.0013016953
#>  [2,] 0.0005352859 0.0005400916 0.0005305775 0.0005949279 0.0006862055
#>  [3,] 0.0003386384 0.0004972456 0.0004905021 0.0005142389 0.0009896564
#>  [4,] 0.0003996396 0.0002754978 0.0002122871 0.0002634868 0.0001066503
#>  [5,] 0.0023786554 0.0020441770 0.0019038091 0.0021427124 0.0016338783
#>  [6,] 0.0014711389 0.0014098101 0.0013855529 0.0015672855 0.0014851498
#>  [7,] 0.0014978441 0.0014193053 0.0013580588 0.0015060021 0.0014294555
#>  [8,] 0.0024916976 0.0022132201 0.0021102402 0.0023786667 0.0019008445
#>  [9,] 0.0020236876 0.0018857533 0.0018616738 0.0020961085 0.0018185670
#> [10,] 0.0018870654 0.0017868104 0.0017641804 0.0019929269 0.0017743286
#>                V37          V38          V39            V4          V40
#>  [1,] 1.262558e-03 0.0011257952 0.0010382332  1.533492e-04 0.0010231918
#>  [2,] 6.983562e-04 0.0005742720 0.0005092456  4.812971e-05 0.0005184072
#>  [3,] 1.213808e-03 0.0006487277 0.0003227141 -1.028739e-04 0.0004863463
#>  [4,] 6.557835e-05 0.0001865173 0.0002981142  1.037549e-04 0.0002787560
#>  [5,] 1.351738e-03 0.0017089085 0.0019929684  4.515109e-04 0.0018095906
#>  [6,] 1.395378e-03 0.0013405058 0.0012913604  2.107623e-04 0.0012116976
#>  [7,] 1.369728e-03 0.0012632488 0.0011935965  1.879733e-04 0.0011602338
#>  [8,] 1.603063e-03 0.0019284148 0.0021538431  4.612998e-04 0.0019375045
#>  [9,] 1.645079e-03 0.0017707722 0.0017867085  3.112626e-04 0.0015868043
#> [10,] 1.611740e-03 0.0016907909 0.0017064158  3.044624e-04 0.0015363886
#>                V41           V42           V43           V44           V45
#>  [1,] 8.281653e-04  0.0007344442  0.0006696242  0.0005356910  0.0005364873
#>  [2,] 3.792845e-04  0.0002630244  0.0001973888  0.0001583116  0.0001219086
#>  [3,] 8.555852e-05 -0.0003289251 -0.0005857357 -0.0005750706 -0.0009031860
#>  [4,] 2.830462e-04  0.0003869623  0.0004875355  0.0004235691  0.0004884000
#>  [5,] 1.759709e-03  0.0019944083  0.0021553408  0.0018345240  0.0020828588
#>  [6,] 1.031784e-03  0.0009686324  0.0009311141  0.0007816866  0.0008501475
#>  [7,] 9.580413e-04  0.0008767894  0.0008259810  0.0006738641  0.0006935900
#>  [8,] 1.871615e-03  0.0020563961  0.0021737666  0.0018483810  0.0021203248
#>  [9,] 1.431884e-03  0.0014443715  0.0014997030  0.0013403369  0.0015926362
#> [10,] 1.362046e-03  0.0013648971  0.0013870863  0.0011918116  0.0013659390
#>                 V46           V47           V48           V49            V5
#>  [1,]  4.664574e-04  0.0003760166  2.634488e-04  1.495941e-04  2.091011e-04
#>  [2,]  9.680963e-05  0.0001080145  7.676338e-05  4.443371e-05  7.215701e-05
#>  [3,] -7.469951e-04 -0.0003080857 -2.120760e-04 -1.339335e-04 -7.972912e-05
#>  [4,]  4.414343e-04  0.0002823451  1.958097e-04  1.191929e-04  1.265377e-04
#>  [5,]  1.774844e-03  0.0011417310  8.014653e-04  4.752201e-04  5.655174e-04
#>  [6,]  7.083278e-04  0.0004983702  3.477638e-04  2.032842e-04  2.742498e-04
#>  [7,]  5.966598e-04  0.0004575056  3.204646e-04  1.835907e-04  2.531209e-04
#>  [8,]  1.782114e-03  0.0011376362  8.031317e-04  4.737229e-04  5.741263e-04
#>  [9,]  1.326406e-03  0.0008213952  5.717411e-04  3.335656e-04  3.778825e-04
#> [10,]  1.132167e-03  0.0007353686  5.165556e-04  3.032437e-04  3.776756e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  6.111344e-05  5.536606e-05  3.601743e-05  2.962251e-05  2.810531e-05
#>  [2,]  2.003322e-05  1.486132e-05  6.332348e-06  9.223153e-06  8.722041e-06
#>  [3,] -3.032261e-05 -4.320205e-05 -5.533249e-05 -1.073401e-05 -2.027832e-05
#>  [4,]  3.269889e-05  3.969534e-05  3.296586e-05  1.391387e-05  1.558790e-05
#>  [5,]  1.577209e-04  1.659310e-04  1.377012e-04  7.324149e-05  7.859712e-05
#>  [6,]  7.913212e-05  7.200047e-05  5.294491e-05  3.678907e-05  3.939966e-05
#>  [7,]  7.264159e-05  6.746831e-05  4.617198e-05  3.492445e-05  3.439950e-05
#>  [8,]  1.631732e-04  1.658927e-04  1.372787e-04  7.501012e-05  8.192190e-05
#>  [9,]  1.253003e-04  1.144304e-04  9.268980e-05  5.130900e-05  6.079319e-05
#> [10,]  1.133132e-04  1.056165e-04  8.382784e-05  5.017728e-05  5.689957e-05
#>                 V55           V56          V57           V58           V59
#>  [1,]  2.198996e-05  2.609228e-05 2.366829e-05  2.462652e-05  2.225438e-05
#>  [2,]  8.946185e-06  9.910294e-06 1.034337e-05  8.079364e-06  5.052013e-06
#>  [3,] -7.676372e-06 -6.476560e-07 1.082238e-05 -1.108767e-05 -2.334989e-05
#>  [4,]  7.200799e-06  1.094833e-05 6.763342e-06  1.348509e-05  1.724410e-05
#>  [5,]  5.122393e-05  5.509988e-05 3.839466e-05  6.414367e-05  7.234422e-05
#>  [6,]  3.079954e-05  3.171737e-05 2.715204e-05  3.250918e-05  3.057149e-05
#>  [7,]  2.618053e-05  3.030645e-05 2.673177e-05  2.962023e-05  2.763490e-05
#>  [8,]  5.604734e-05  5.752405e-05 4.094240e-05  6.638125e-05  7.252398e-05
#>  [9,]  4.826434e-05  4.293589e-05 3.192197e-05  4.752127e-05  4.748431e-05
#> [10,]  4.377083e-05  4.182403e-05 3.297502e-05  4.602916e-05  4.564449e-05
#>                  V6           V60           V7            V8            V9
#>  [1,]  3.706923e-04  2.049468e-05 4.587514e-04  0.0006114390  0.0006925100
#>  [2,]  1.318524e-04  6.168522e-06 1.850644e-04  0.0002393299  0.0002392656
#>  [3,] -6.247623e-05 -7.123950e-06 6.362353e-05 -0.0001148673 -0.0003903348
#>  [4,]  2.026910e-04  1.159914e-05 2.004065e-04  0.0003389953  0.0004795718
#>  [5,]  8.910590e-04  5.078727e-05 9.388443e-04  0.0014639535  0.0019557779
#>  [6,]  4.600608e-04  2.470776e-05 5.425404e-04  0.0007794974  0.0009318186
#>  [7,]  4.384806e-04  2.418781e-05 5.322019e-04  0.0007290868  0.0008519606
#>  [8,]  9.043338e-04  5.077057e-05 9.616792e-04  0.0015004220  0.0019749203
#>  [9,]  6.282337e-04  3.460816e-05 7.185972e-04  0.0011470131  0.0014801571
#> [10,]  6.177230e-04  3.401028e-05 7.047607e-04  0.0010745816  0.0013492546
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]         [,4]        [,5]        [,6]
#> [1,] -0.02969714 -0.08733273 -0.07016123 -0.004915985 -0.03744777 -0.07201086
#> [2,]  0.02969714  0.08733273  0.07016123  0.004915985  0.03744777  0.07201086
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.03492924 -0.0479568 -0.0712134 -0.06741623
#> [2,]  0.03492924  0.0479568  0.0712134  0.06741623
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.08401422  0.74810641  0.53399868 -0.08135802  0.15104896  0.60044781
#>  [7]  0.50563452  0.47082178  0.54138919  0.51347366
#> 
#> $B[[2]]
#> [1] 0.07335653 0.00174087
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.003366372 0.001400018 0.001159984 0.001080170 0.005899975 0.003682934
#>  [7] 0.003792972 0.006085195 0.004465818 0.004550755
#> 
#> $vB[[2]]
#> [1] -0.09498707  0.09498707
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#>  [2,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [3,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#>  [4,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#>  [5,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#>  [6,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [7,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#>  [8,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#>  [9,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [10,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [11,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [12,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [13,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [14,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [15,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [16,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [17,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#> [18,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [19,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [20,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [21,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [22,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [23,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [24,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [25,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [26,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [27,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [28,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [29,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [30,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [31,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [32,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [33,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [34,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [35,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [36,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [37,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [38,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [39,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [40,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#>  [2,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [3,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#>  [4,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#>  [5,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#>  [6,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [7,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#>  [8,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#>  [9,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [10,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [11,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [12,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [13,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [14,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [15,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [16,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [17,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#> [18,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [19,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [20,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [21,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [22,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [23,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [24,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [25,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [26,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [27,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [28,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [29,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [30,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [31,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [32,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [33,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [34,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [35,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [36,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [37,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [38,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [39,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [40,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#>  [2,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [3,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#>  [4,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#>  [5,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#>  [6,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [7,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#>  [8,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#>  [9,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [10,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [11,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [12,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [13,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [14,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [15,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [16,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [17,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#> [18,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [19,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [20,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [21,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [22,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [23,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [24,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [25,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [26,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [27,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [28,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [29,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [30,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [31,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [32,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [33,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [34,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [35,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [36,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [37,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [38,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [39,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [40,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#>  [2,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [3,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#>  [4,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#>  [5,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#>  [6,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [7,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#>  [8,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#>  [9,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [10,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [11,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [12,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [13,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [14,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [15,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [16,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [17,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#> [18,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [19,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [20,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [21,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [22,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [23,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [24,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [25,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [26,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [27,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [28,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [29,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [30,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [31,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [32,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [33,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [34,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [35,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [36,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [37,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [38,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [39,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [40,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#>  [2,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [3,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#>  [4,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#>  [5,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#>  [6,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [7,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#>  [8,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#>  [9,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [10,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [11,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [12,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [13,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [14,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [15,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [16,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [17,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#> [18,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [19,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [20,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [21,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [22,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [23,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [24,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [25,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [26,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [27,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [28,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [29,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [30,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [31,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [32,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [33,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [34,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [35,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [36,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [37,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [38,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [39,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [40,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#>  [2,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [3,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#>  [4,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#>  [5,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#>  [6,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [7,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#>  [8,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#>  [9,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [10,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [11,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [12,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [13,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [14,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [15,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [16,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [17,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#> [18,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [19,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [20,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [21,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [22,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [23,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [24,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [25,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [26,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [27,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [28,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [29,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [30,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [31,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [32,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [33,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [34,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [35,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [36,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [37,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [38,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [39,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [40,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5193662 0.9366924 0.8360612 0.20705748 0.5507984 0.8280786 0.5663940
#>  [2,] 0.4701242 0.9359798 0.8416840 0.13696179 0.5392274 0.8188244 0.5250849
#>  [3,] 0.5122179 0.9338354 0.8503800 0.21323046 0.5804438 0.8378097 0.5707825
#>  [4,] 0.4823399 0.9027588 0.7427026 0.17015407 0.4849392 0.7901802 0.4978962
#>  [5,] 0.4526475 0.9121497 0.7929066 0.19856670 0.5359561 0.8040398 0.5286156
#>  [6,] 0.4852612 0.9490664 0.8234212 0.15790895 0.5373873 0.8606327 0.5263248
#>  [7,] 0.4307831 0.9129579 0.7667187 0.10507630 0.4478253 0.7658396 0.4409949
#>  [8,] 0.4815167 0.9043891 0.7831952 0.15635693 0.5080870 0.7792808 0.5290206
#>  [9,] 0.3013064 0.8840542 0.5625898 0.08041568 0.2939619 0.7057048 0.3606856
#> [10,] 0.4496117 0.9259818 0.8055408 0.12088545 0.5021796 0.7896010 0.4967619
#> [11,] 0.4569615 0.9220839 0.8285525 0.18030946 0.5640831 0.8185367 0.5266721
#> [12,] 0.4653851 0.9234619 0.7754776 0.14160347 0.5068339 0.7985481 0.4860762
#> [13,] 0.2991171 0.9010165 0.6386022 0.07374582 0.3707828 0.7400237 0.3349086
#> [14,] 0.3960551 0.8596578 0.6181079 0.11590315 0.4210678 0.6985429 0.4247540
#> [15,] 0.4645479 0.9395694 0.8553254 0.15330961 0.5458714 0.8296276 0.5452501
#> [16,] 0.4890153 0.9424598 0.8638091 0.17867039 0.5640639 0.8373623 0.5617417
#> [17,] 0.4209571 0.9083588 0.7457362 0.10642919 0.4718828 0.7598439 0.4936312
#> [18,] 0.4638926 0.9021588 0.7646931 0.14443879 0.4913878 0.7600213 0.5344152
#> [19,] 0.4457185 0.8865164 0.6788574 0.11404038 0.4341641 0.7466462 0.4392820
#> [20,] 0.4463043 0.9079685 0.7610007 0.20730557 0.5009834 0.7912063 0.5410270
#> [21,] 0.4891637 0.9378478 0.8494952 0.18695220 0.5840023 0.8435445 0.5654014
#> [22,] 0.4344832 0.9333540 0.7504317 0.09210988 0.4495429 0.8057726 0.4587115
#> [23,] 0.3722592 0.9076335 0.6836315 0.09963515 0.4130041 0.7690041 0.3538084
#> [24,] 0.4854227 0.9418571 0.8392064 0.17672778 0.5625335 0.8431670 0.5403386
#> [25,] 0.4698142 0.9240907 0.8270123 0.14622960 0.5235541 0.7921168 0.5339190
#> [26,] 0.4379575 0.9115241 0.7461494 0.10439942 0.4418885 0.7667513 0.4641651
#> [27,] 0.4194118 0.9415509 0.7987133 0.10030722 0.4628284 0.8179991 0.4800722
#> [28,] 0.4699971 0.9414838 0.8413842 0.12677094 0.5211773 0.8290468 0.5157366
#> [29,] 0.4124554 0.9170762 0.7578643 0.09839014 0.4706160 0.7665888 0.4765672
#> [30,] 0.3472930 0.9103368 0.6333059 0.09784452 0.3932715 0.7903494 0.3679434
#> [31,] 0.4566018 0.9423878 0.7605844 0.09653666 0.4791747 0.8295594 0.4910401
#> [32,] 0.5194714 0.9584154 0.8720554 0.17140706 0.5851872 0.8733578 0.5733648
#> [33,] 0.4324923 0.9261434 0.7590394 0.15905941 0.5042478 0.8111389 0.4972784
#> [34,] 0.4440948 0.9294717 0.7586466 0.11820363 0.4828329 0.8088842 0.4813725
#> [35,] 0.3377308 0.9126911 0.6340374 0.08719626 0.3880986 0.7890283 0.3497677
#> [36,] 0.5298637 0.9301000 0.8518542 0.17276818 0.5609130 0.8137972 0.5820326
#> [37,] 0.3753988 0.9069634 0.6787509 0.10837673 0.4090613 0.7766436 0.4496935
#> [38,] 0.4938740 0.9464335 0.8696460 0.17667439 0.5711885 0.8437302 0.5721706
#> [39,] 0.3925901 0.8973389 0.5953369 0.14743422 0.4114423 0.7768353 0.3990991
#> [40,] 0.3347264 0.8563405 0.5827789 0.11150664 0.3220699 0.6841290 0.3903995
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6798688 0.8430288 0.7971678
#>  [2,] 0.6624563 0.8840743 0.8165955
#>  [3,] 0.7132640 0.8663556 0.8286417
#>  [4,] 0.5755220 0.7856905 0.7341758
#>  [5,] 0.6627467 0.8022060 0.7846529
#>  [6,] 0.6466487 0.8272329 0.8049360
#>  [7,] 0.5032469 0.8386063 0.7324689
#>  [8,] 0.6172809 0.8526324 0.7722544
#>  [9,] 0.3343278 0.5564477 0.5005460
#> [10,] 0.5950093 0.8750245 0.7751226
#> [11,] 0.6855283 0.8330469 0.8071401
#> [12,] 0.6046325 0.8225370 0.7405337
#> [13,] 0.4410267 0.6933672 0.6966397
#> [14,] 0.4962745 0.7589899 0.6567638
#> [15,] 0.6927969 0.8689241 0.8184960
#> [16,] 0.7171775 0.8726319 0.8331777
#> [17,] 0.5772026 0.8481580 0.7591243
#> [18,] 0.6050562 0.8333500 0.7387661
#> [19,] 0.4708779 0.7871785 0.6769169
#> [20,] 0.6662594 0.7697178 0.7599434
#> [21,] 0.7224887 0.8656367 0.8353436
#> [22,] 0.5036531 0.8399589 0.7557737
#> [23,] 0.4458470 0.7385690 0.6838550
#> [24,] 0.6974049 0.8502665 0.8254499
#> [25,] 0.6528260 0.8701250 0.7863607
#> [26,] 0.5264080 0.8414433 0.7373198
#> [27,] 0.5667072 0.8173427 0.7535871
#> [28,] 0.6551841 0.8825998 0.8149058
#> [29,] 0.5803944 0.8464041 0.7604341
#> [30,] 0.4566153 0.6539887 0.6816020
#> [31,] 0.5897112 0.8387140 0.7816830
#> [32,] 0.7147248 0.8809949 0.8373896
#> [33,] 0.6011264 0.7712177 0.7545582
#> [34,] 0.5661104 0.8161942 0.7346658
#> [35,] 0.4551338 0.6600235 0.6933223
#> [36,] 0.6905948 0.9023478 0.8233376
#> [37,] 0.4765595 0.7494426 0.6798894
#> [38,] 0.7205700 0.8793050 0.8340621
#> [39,] 0.4503791 0.6290344 0.6436125
#> [40,] 0.3658317 0.5680218 0.5281685
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.3508835 0.6491165
#>  [2,] 0.3529454 0.6470546
#>  [3,] 0.3470895 0.6529105
#>  [4,] 0.3617350 0.6382650
#>  [5,] 0.3548534 0.6451466
#>  [6,] 0.3539483 0.6460517
#>  [7,] 0.3674854 0.6325146
#>  [8,] 0.3570592 0.6429408
#>  [9,] 0.3955532 0.6044468
#> [10,] 0.3588861 0.6411139
#> [11,] 0.3521697 0.6478303
#> [12,] 0.3598434 0.6401566
#> [13,] 0.3801495 0.6198505
#> [14,] 0.3726143 0.6273857
#> [15,] 0.3515466 0.6484534
#> [16,] 0.3485245 0.6514755
#> [17,] 0.3621740 0.6378260
#> [18,] 0.3600124 0.6399876
#> [19,] 0.3709256 0.6290744
#> [20,] 0.3570813 0.6429187
#> [21,] 0.3472608 0.6527392
#> [22,] 0.3656834 0.6343166
#> [23,] 0.3758321 0.6241679
#> [24,] 0.3500330 0.6499670
#> [25,] 0.3550346 0.6449654
#> [26,] 0.3659313 0.6340687
#> [27,] 0.3640294 0.6359706
#> [28,] 0.3540544 0.6459456
#> [29,] 0.3626646 0.6373354
#> [30,] 0.3779074 0.6220926
#> [31,] 0.3599013 0.6400987
#> [32,] 0.3466148 0.6533852
#> [33,] 0.3606793 0.6393207
#> [34,] 0.3627306 0.6372694
#> [35,] 0.3782514 0.6217486
#> [36,] 0.3485756 0.6514244
#> [37,] 0.3728966 0.6271034
#> [38,] 0.3477467 0.6522533
#> [39,] 0.3769316 0.6230684
#> [40,] 0.3916405 0.6083595
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]     [,2]      [,3]      [,4]         [,5]      [,6]
#>  [1,]  0.07750360 2.694350 1.6292084 -1.342754  0.203897028 1.5720710
#>  [2,] -0.11964589 2.682395 1.6708117 -1.840757  0.157232714 1.5084028
#>  [3,]  0.04888126 2.647155 1.7375845 -1.305562  0.324595479 1.6420204
#>  [4,] -0.07066974 2.228261 1.0600630 -1.584536 -0.060261335 1.3260118
#>  [5,] -0.18997942 2.340170 1.3425352 -1.395277  0.144073286 1.4117371
#>  [6,] -0.05897225 2.924955 1.5397008 -1.673870  0.149828889 1.8205553
#>  [7,] -0.27865679 2.350298 1.1898751 -2.142052 -0.209461462 1.1849665
#>  [8,] -0.07396676 2.246973 1.2843847 -1.685588  0.032350920 1.2614800
#>  [9,] -0.84108466 2.031395 0.2516792 -2.436713 -0.876218982 0.8746139
#> [10,] -0.20223989 2.526544 1.4212914 -1.984072  0.008718297 1.3225218
#> [11,] -0.17258113 2.471004 1.5754034 -1.514252  0.257749869 1.5064645
#> [12,] -0.13868128 2.490342 1.2395039 -1.802035  0.027337341 1.3772447
#> [13,] -0.85150575 2.208570 0.5693025 -2.530524 -0.528860022 1.0460919
#> [14,] -0.42192951 1.812451 0.4815250 -2.031812 -0.318391601 0.8403691
#> [15,] -0.14204685 2.743926 1.7769952 -1.708876  0.184002952 1.5829901
#> [16,] -0.04394597 2.796010 1.8472943 -1.525382  0.257671977 1.6387317
#> [17,] -0.31884552 2.293758 1.0759998 -2.127746 -0.112587775 1.1518238
#> [18,] -0.14468157 2.221445 1.1785837 -1.778902 -0.034452166 1.1527961
#> [19,] -0.21798499 2.055641 0.7485256 -2.050119 -0.264881450 1.0808046
#> [20,] -0.21561438 2.289079 1.1581739 -1.341244  0.003933422 1.3322118
#> [21,] -0.04335203 2.714001 1.7306474 -1.469937  0.339225033 1.6848413
#> [22,] -0.26358261 2.639390 1.1009162 -2.288141 -0.202517553 1.4227722
#> [23,] -0.52253681 2.285076 0.7705115 -2.201285 -0.351560473 1.2026961
#> [24,] -0.05832576 2.784949 1.6523355 -1.538676  0.251450447 1.6819831
#> [25,] -0.12089036 2.499271 1.5645989 -1.764484  0.094286342 1.3377325
#> [26,] -0.24945544 2.332387 1.0781802 -2.149270 -0.233501315 1.1900572
#> [27,] -0.32518858 2.779371 1.3782715 -2.193816 -0.148961328 1.5028499
#> [28,] -0.12015614 2.778153 1.6685638 -1.929816  0.084759747 1.5788866
#> [29,] -0.35382394 2.403268 1.1410063 -2.215241 -0.117671728 1.1891489
#> [30,] -0.63095958 2.317754 0.5464258 -2.221407 -0.433581165 1.3270328
#> [31,] -0.17403078 2.794682 1.1558861 -2.236313 -0.083349435 1.5825075
#> [32,]  0.07792486 3.137552 1.9192552 -1.575688  0.344104666 1.9309799
#> [33,] -0.27168998 2.528904 1.1474200 -1.665243  0.016991476 1.4574274
#> [34,] -0.22455961 2.578603 1.1452739 -2.009552 -0.068695589 1.4427764
#> [35,] -0.67342321 2.346945 0.5495768 -2.348359 -0.455311724 1.3190783
#> [36,]  0.11959697 2.588226 1.7492184 -1.566134  0.244868207 1.4748750
#> [37,] -0.50912445 2.277109 0.7480372 -2.107430 -0.367847369 1.2462129
#> [38,] -0.02450507 2.871777 1.8978325 -1.539043  0.286702013 1.6862489
#> [39,] -0.43643772 2.168000 0.3860726 -1.754868 -0.358006122 1.2473183
#> [40,] -0.68688469 1.785222 0.3341914 -2.075443 -0.744275430 0.7728124
#>               [,7]        [,8]      [,9]       [,10]
#>  [1,]  0.267153787  0.75316893 1.6809388 1.368686213
#>  [2,]  0.100423807  0.67425975 2.0315915 1.493449556
#>  [3,]  0.285044243  0.91128976 1.8691130 1.576030883
#>  [4,] -0.008415442  0.30441744 1.2991419 1.015913167
#>  [5,]  0.114587763  0.67555861 1.4001392 1.292990474
#>  [6,]  0.105396477  0.60434027 1.5661417 1.417434898
#>  [7,] -0.237125384  0.01298792 1.6478949 1.007184987
#>  [8,]  0.116213191  0.47802292 1.7553985 1.221084777
#>  [9,] -0.572389528 -0.68867545 0.2267574 0.002184163
#> [10,] -0.012952397  0.38471298 1.9461343 1.237466007
#> [11,]  0.106789937  0.77929567 1.6073770 1.431533164
#> [12,] -0.055709667  0.42480526 1.5336315 1.048744181
#> [13,] -0.686066966 -0.23699634 0.8159086 0.831346952
#> [14,] -0.303287735 -0.01490221 1.1471498 0.648905740
#> [15,]  0.181496834  0.81322784 1.8914796 1.506190562
#> [16,]  0.248233768  0.93050366 1.9244323 1.608317419
#> [17,] -0.025476687  0.31130007 1.7202264 1.147884657
#> [18,]  0.137878990  0.42657808 1.6095577 1.039564968
#> [19,] -0.244076450 -0.11662027 1.3080011 0.739639086
#> [20,]  0.164477914  0.69131507 1.2067187 1.152369163
#> [21,]  0.263113216  0.95684033 1.8629183 1.623981968
#> [22,] -0.165531046  0.01461254 1.6579223 1.129646532
#> [23,] -0.602340515 -0.21746492 1.0385440 0.771545090
#> [24,]  0.161705894  0.83497061 1.7366929 1.553716875
#> [25,]  0.135884518  0.63148458 1.9020648 1.303126256
#> [26,] -0.143585629  0.10573046 1.6690065 1.032084223
#> [27,] -0.079753417  0.26842889 1.4984465 1.117836233
#> [28,]  0.062967019  0.64190574 2.0172830 1.482207368
#> [29,] -0.093799817  0.32439248 1.7066716 1.155060736
#> [30,] -0.541049833 -0.17397644 0.6366186 0.761143513
#> [31,] -0.035843578  0.36277178 1.6486906 1.275501184
#> [32,]  0.295592718  0.91844314 2.0018858 1.638932447
#> [33,] -0.010886640  0.41016047 1.2151998 1.123072569
#> [34,] -0.074544647  0.26599882 1.4907725 1.018425719
#> [35,] -0.620060487 -0.17994898 0.6633990 0.815697737
#> [36,]  0.331123067  0.80290153 2.2235875 1.539126090
#> [37,] -0.201909195 -0.09383087 1.0956415 0.753263618
#> [38,]  0.290712823  0.94729066 1.9858656 1.614694178
#> [39,] -0.409220359 -0.19913900 0.5280764 0.591078092
#> [40,] -0.445633182 -0.55014116 0.2737847 0.112793412
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.2705339 0.3446242
#>  [2,] -0.2639227 0.3421945
#>  [3,] -0.2795378 0.3523197
#>  [4,] -0.2467878 0.3210539
#>  [5,] -0.2622861 0.3354868
#>  [6,] -0.2613340 0.3403948
#>  [7,] -0.2325660 0.3104535
#>  [8,] -0.2576233 0.3305279
#>  [9,] -0.1730450 0.2509834
#> [10,] -0.2521202 0.3280820
#> [11,] -0.2670401 0.3424754
#> [12,] -0.2497797 0.3262643
#> [13,] -0.2025830 0.2863307
#> [14,] -0.2259324 0.2950854
#> [15,] -0.2668707 0.3453770
#> [16,] -0.2740467 0.3514845
#> [17,] -0.2460393 0.3199013
#> [18,] -0.2517768 0.3235334
#> [19,] -0.2281103 0.3001377
#> [20,] -0.2573771 0.3306777
#> [21,] -0.2783045 0.3527969
#> [22,] -0.2367913 0.3139890
#> [23,] -0.2133261 0.2939509
#> [24,] -0.2706814 0.3482126
#> [25,] -0.2603881 0.3365933
#> [26,] -0.2365006 0.3132113
#> [27,] -0.2379588 0.3199591
#> [28,] -0.2601153 0.3411493
#> [29,] -0.2434579 0.3203598
#> [30,] -0.2077592 0.2906807
#> [31,] -0.2478548 0.3279378
#> [32,] -0.2784042 0.3555485
#> [33,] -0.2485035 0.3239134
#> [34,] -0.2436667 0.3198654
#> [35,] -0.2059060 0.2910708
#> [36,] -0.2757794 0.3495269
#> [37,] -0.2234216 0.2963886
#> [38,] -0.2761723 0.3527862
#> [39,] -0.2141028 0.2884898
#> [40,] -0.1821258 0.2582959
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.3508835  0.3508835
#>  [2,]  0.6470546 -0.6470546
#>  [3,] -0.3470895  0.3470895
#>  [4,] -0.3617350  0.3617350
#>  [5,]  0.6451466 -0.6451466
#>  [6,] -0.3539483  0.3539483
#>  [7,]  0.6325146 -0.6325146
#>  [8,] -0.3570592  0.3570592
#>  [9,]  0.6044468 -0.6044468
#> [10,]  0.6411139 -0.6411139
#> [11,]  0.6478303 -0.6478303
#> [12,]  0.6401566 -0.6401566
#> [13,]  0.6198505 -0.6198505
#> [14,] -0.3726143  0.3726143
#> [15,]  0.6484534 -0.6484534
#> [16,]  0.6514755 -0.6514755
#> [17,] -0.3621740  0.3621740
#> [18,]  0.6399876 -0.6399876
#> [19,] -0.3709256  0.3709256
#> [20,] -0.3570813  0.3570813
#> [21,]  0.6527392 -0.6527392
#> [22,] -0.3656834  0.3656834
#> [23,]  0.6241679 -0.6241679
#> [24,] -0.3500330  0.3500330
#> [25,]  0.6449654 -0.6449654
#> [26,]  0.6340687 -0.6340687
#> [27,]  0.6359706 -0.6359706
#> [28,]  0.6459456 -0.6459456
#> [29,] -0.3626646  0.3626646
#> [30,]  0.6220926 -0.6220926
#> [31,]  0.6400987 -0.6400987
#> [32,] -0.3466148  0.3466148
#> [33,] -0.3606793  0.3606793
#> [34,] -0.3627306  0.3627306
#> [35,]  0.6217486 -0.6217486
#> [36,] -0.3485756  0.3485756
#> [37,] -0.3728966  0.3728966
#> [38,]  0.6522533 -0.6522533
#> [39,] -0.3769316  0.3769316
#> [40,]  0.6083595 -0.6083595
#> 
#> $L
#> [1] 0.7041269 0.7730391 0.7939841 0.6893067 0.7589414 0.7576792
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4202899