Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.1840600 -0.06608869 -0.12549977 -0.082453944 -0.16168301 -0.153863941
#>  [2,] -0.3378555 -0.11212614 -0.17894594 -0.206840504 -0.06702972 -0.016270522
#>  [3,] -0.2671518 -0.07429475 -0.01999422  0.005199330 -0.06801791  0.009217737
#>  [4,] -0.2714283 -0.14099093 -0.19222829 -0.212706591 -0.10234669 -0.205009766
#>  [5,] -0.4012064 -0.07427741 -0.11034037 -0.021505871 -0.13851173 -0.118133949
#>  [6,] -0.3667456 -0.20724989 -0.14207144 -0.080934673 -0.12938383 -0.109706779
#>  [7,] -0.1656782 -0.08692457 -0.04741861 -0.161214487 -0.13178151 -0.121580876
#>  [8,] -0.2806684 -0.06367933 -0.03725899  0.008863362  0.03875403  0.172366867
#>  [9,] -0.3709396 -0.16154339 -0.04777942 -0.067097126 -0.15223621 -0.175294058
#> [10,] -0.2889428 -0.08183689 -0.14016734 -0.162760837  0.01232321  0.025202299
#>                V15          V16         V17           V18         V19
#>  [1,] -0.096061546  0.018902845 -0.08460314  0.0130273418  0.08697138
#>  [2,] -0.161369335 -0.187365637 -0.03695352 -0.1060210499 -0.15632427
#>  [3,]  0.071308909  0.041090464 -0.03721526  0.1293899275  0.03213869
#>  [4,] -0.121579045 -0.093163793 -0.01483519 -0.1602124655  0.07095391
#>  [5,] -0.007557335 -0.117055856 -0.01168752  0.0576920410 -0.07810946
#>  [6,] -0.102364159 -0.087058528  0.03480958  0.0419544475 -0.07184919
#>  [7,] -0.129570589 -0.067353218 -0.10471144 -0.0920413979 -0.04310691
#>  [8,]  0.021344842  0.077937349  0.06561503  0.0942103666  0.11547820
#>  [9,] -0.157494573 -0.020920304 -0.02955928 -0.1041306118 -0.09728454
#> [10,] -0.062901462  0.008055583  0.05410495  0.0002103555  0.09960250
#>               V2          V20         V21         V22         V23        V24
#>  [1,] -0.3085166 -0.013182834  0.02833862 0.062926883  0.15047026 0.15579325
#>  [2,] -0.3757513 -0.160903465 -0.08585281 0.094637529  0.06598528 0.17023589
#>  [3,] -0.3908248  0.075951981  0.01577943 0.116949535  0.23886416 0.23286480
#>  [4,] -0.2733635  0.036258514  0.07107252 0.014113205  0.15161777 0.13946962
#>  [5,] -0.3068493 -0.038410647  0.06552219 0.174381170  0.23632637 0.27118228
#>  [6,] -0.2085328  0.079997538 -0.03697755 0.089608564  0.10163059 0.03581430
#>  [7,] -0.2373663 -0.113331347  0.03034815 0.003875728  0.02528015 0.02353099
#>  [8,] -0.3022751  0.238245183  0.25567312 0.188554188  0.07909883 0.11496549
#>  [9,] -0.3465295  0.009160016 -0.04568004 0.022120426 -0.02846390 0.08036688
#> [10,] -0.2248583  0.022748997  0.10951386 0.072402160  0.25369849 0.09812598
#>                V25        V26         V27        V28         V29         V3
#>  [1,]  0.241013888 0.18121943  0.09275110 0.12481792 0.051545600 -0.1648737
#>  [2,]  0.135201721 0.19628050  0.12649417 0.18449046 0.155935061 -0.2503953
#>  [3,]  0.153175652 0.22568534  0.17667911 0.15949707 0.150118322 -0.2471044
#>  [4,]  0.196787917 0.14499643  0.19186329 0.16146443 0.062285608 -0.3716898
#>  [5,]  0.249609549 0.22345645  0.19503412 0.24171716 0.221833725 -0.2656518
#>  [6,]  0.113864215 0.08800223 -0.01975148 0.05932888 0.008547247 -0.2928620
#>  [7,]  0.052547667 0.05314294  0.05007069 0.23122164 0.117908959 -0.3132985
#>  [8,]  0.148636066 0.17038394  0.29703782 0.20745749 0.058427704 -0.3201849
#>  [9,] -0.003182448 0.01405569  0.02324200 0.02764276 0.042890875 -0.2056113
#> [10,]  0.123485423 0.27309169  0.20647227 0.08989993 0.115948902 -0.3144232
#>               V30          V31           V32         V33          V34
#>  [1,] -0.02012743 -0.032256401 -0.0095478608 -0.05829802 -0.035183021
#>  [2,]  0.03373215  0.001009969  0.1145003821  0.03197938  0.029532032
#>  [3,]  0.14886738 -0.009274161  0.0139911464 -0.04779336  0.005521781
#>  [4,] -0.03052840 -0.021811309  0.0002315304 -0.01428159  0.037676754
#>  [5,]  0.21716392  0.063822927  0.0297714827 -0.04544498 -0.042094397
#>  [6,] -0.04314060  0.052731055 -0.0741489005  0.07932124  0.025284044
#>  [7,]  0.14793026  0.086645743  0.0638473891  0.19006518  0.132998088
#>  [8,]  0.10784177  0.066875440  0.0408664717 -0.01257883  0.035256893
#>  [9,] -0.04654281 -0.017598005 -0.0420763096 -0.06814550 -0.062293634
#> [10,]  0.08091591  0.169252536  0.1311874724  0.02984058 -0.078125078
#>                 V35         V36         V37          V38           V39
#>  [1,] -0.0416759890 -0.03984658 -0.15782496 -0.163507903 -0.2102186560
#>  [2,]  0.0279125479  0.03558321 -0.14218585 -0.185619501 -0.0706884201
#>  [3,] -0.1107273244 -0.11050744 -0.11552385 -0.186935771  0.0006169502
#>  [4,] -0.0005913766  0.02367054  0.04506898 -0.055851128 -0.1147243427
#>  [5,]  0.0797225371 -0.08668590 -0.05652403 -0.059115367 -0.1654021160
#>  [6,] -0.0367333146 -0.01356295 -0.07787767  0.058386704 -0.0183883299
#>  [7,]  0.1133821500 -0.05147462  0.02656860  0.006321672  0.0537070372
#>  [8,] -0.0990397890  0.02157751  0.03388509 -0.020670510  0.0728673280
#>  [9,] -0.1872185756 -0.15824473 -0.17850240 -0.051857174 -0.1552357113
#> [10,]  0.0594043275  0.01613687  0.06393404 -0.034777195 -0.0555055466
#>               V4         V40          V41          V42         V43         V44
#>  [1,] -0.1511711 -0.14006865  0.007183207  0.007002509 -0.06611886 -0.19859316
#>  [2,] -0.3498832  0.01415402  0.002713782 -0.074710851 -0.09479803 -0.03188781
#>  [3,] -0.3490725  0.03234222 -0.118791839 -0.037192246 -0.13899466 -0.03536789
#>  [4,] -0.3346712 -0.14115022 -0.020288000 -0.133283465 -0.09298598 -0.18561980
#>  [5,] -0.2762801 -0.15747124 -0.079194181  0.026216109 -0.00991698 -0.02843272
#>  [6,] -0.2676285 -0.04730611  0.022202853 -0.161804088 -0.04850476 -0.15473232
#>  [7,] -0.2595053  0.03656570 -0.046239240  0.028594492  0.03407036 -0.02646750
#>  [8,] -0.1866400 -0.06420819  0.110052008  0.108653974 -0.04186552 -0.11386462
#>  [9,] -0.1808377 -0.08598488 -0.196908240 -0.128917449 -0.17873810 -0.14558683
#> [10,] -0.3003133  0.01366471 -0.075037515 -0.052340008 -0.12923165 -0.12862434
#>                V45          V46         V47        V48        V49         V5
#>  [1,] -0.142792060 -0.139180208 -0.16975894 -0.2514948 -0.2099295 -0.2001329
#>  [2,] -0.175173847 -0.095821751 -0.17130465 -0.2401963 -0.3314972 -0.2718319
#>  [3,] -0.221033965 -0.151130771 -0.09912831 -0.2579976 -0.2298471 -0.2863973
#>  [4,] -0.070012456 -0.099047533 -0.15795327 -0.2306205 -0.2424949 -0.2503813
#>  [5,] -0.113447919 -0.138834613 -0.26548158 -0.2327047 -0.3473440 -0.3289822
#>  [6,] -0.036934897 -0.122826045 -0.15896363 -0.1495163 -0.1322922 -0.2198899
#>  [7,]  0.002857704  0.008335815 -0.15989257 -0.2234428 -0.2500639 -0.1329625
#>  [8,] -0.108023623 -0.155880041 -0.02899021 -0.1873895 -0.2871217 -0.2606770
#>  [9,] -0.097373454 -0.152857657 -0.27654086 -0.2543854 -0.3566783 -0.3127163
#> [10,]  0.006895290 -0.040237423 -0.12876588 -0.2301603 -0.1592797 -0.2450589
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2790935 -0.3631838 -0.3053524 -0.3455821 -0.3426556 -0.2719939
#>  [2,] -0.3516896 -0.3426907 -0.4013681 -0.2956149 -0.4184614 -0.3617152
#>  [3,] -0.3464064 -0.3649768 -0.2834432 -0.3681658 -0.3952439 -0.3872301
#>  [4,] -0.2741079 -0.4251072 -0.4279814 -0.3426331 -0.3213913 -0.4350588
#>  [5,] -0.3269355 -0.3162381 -0.2797624 -0.3087165 -0.3656862 -0.2989043
#>  [6,] -0.3349044 -0.3075707 -0.3306785 -0.3273684 -0.3589675 -0.2721528
#>  [7,] -0.2857878 -0.3640210 -0.3185265 -0.2303246 -0.3405254 -0.3462787
#>  [8,] -0.3832581 -0.3304230 -0.2452564 -0.3905048 -0.2815698 -0.3964274
#>  [9,] -0.3730790 -0.2448849 -0.3404572 -0.3203748 -0.2514069 -0.3094394
#> [10,] -0.2265327 -0.2328368 -0.3097013 -0.2540012 -0.2581404 -0.2740370
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3105363 -0.2676023 -0.3029397 -0.3646165 -0.2292705 -0.2666487
#>  [2,] -0.3379239 -0.2809252 -0.3135805 -0.2564538 -0.1711371 -0.4372494
#>  [3,] -0.3753862 -0.4305906 -0.4291606 -0.3290581 -0.3159056 -0.4523506
#>  [4,] -0.3937098 -0.3893610 -0.4299384 -0.4218415 -0.1860172 -0.2898331
#>  [5,] -0.3012110 -0.4134020 -0.3067341 -0.4329373 -0.2042407 -0.3829891
#>  [6,] -0.3785443 -0.3082533 -0.3922788 -0.3131089 -0.2050919 -0.2172641
#>  [7,] -0.3281098 -0.3308692 -0.1931051 -0.2858483 -0.2039022 -0.3092952
#>  [8,] -0.3745782 -0.3417362 -0.3446690 -0.3318404 -0.1102172 -0.3441300
#>  [9,] -0.3872423 -0.2418120 -0.3218518 -0.2882075 -0.2293671 -0.3772697
#> [10,] -0.3669058 -0.3414222 -0.3990297 -0.3623528 -0.1337051 -0.4192937
#>               V7          V8           V9
#>  [1,] -0.1363259 -0.21115284 -0.122653448
#>  [2,] -0.1092445 -0.21574884 -0.171503266
#>  [3,] -0.2258595 -0.16551884 -0.104653209
#>  [4,] -0.2158685 -0.11697835 -0.184716582
#>  [5,] -0.1981550 -0.22867630 -0.184844078
#>  [6,] -0.1212774 -0.10932096 -0.120919969
#>  [7,] -0.1836765 -0.13841306 -0.073690579
#>  [8,] -0.1302488 -0.03980341  0.003978526
#>  [9,] -0.2999922 -0.16003121 -0.149406753
#> [10,] -0.2168254 -0.15425907 -0.125525142
#> 
#> $W[[2]]
#>             [,1]        [,2]       [,3]          [,4]        [,5]        [,6]
#> [1,] -0.08383047 -0.02931670 0.05535055 -0.0473470722  0.06882011 -0.01322015
#> [2,]  0.06281147 -0.02158582 0.03264641  0.0006994709 -0.05231892  0.03981196
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01306666  0.04330923 -0.07554973 -0.09112623
#> [2,] -0.04688807 -0.12268469  0.02932621 -0.03370912
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.542897e-04  1.751218e-03  2.139154e-03  2.100042e-03  2.164684e-03
#>  [2,]  5.570281e-05  4.283575e-04  4.146931e-04  4.947899e-04  5.432912e-04
#>  [3,]  8.145283e-05  6.079893e-04  6.304787e-04  6.467245e-04  6.873544e-04
#>  [4,]  1.058970e-04  7.700711e-04  8.521817e-04  9.024036e-04  9.550478e-04
#>  [5,] -2.264842e-05  6.071561e-05 -9.323170e-05  2.003720e-05  1.412990e-05
#>  [6,]  1.113026e-04  8.131103e-04  9.132578e-04  9.573819e-04  1.006019e-03
#>  [7,] -2.477159e-06  8.653667e-05 -4.044148e-05  4.666059e-05  9.154491e-05
#>  [8,] -1.086688e-05 -7.168862e-06 -1.091460e-04 -4.059273e-05 -2.777190e-05
#>  [9,]  1.043513e-04  8.448875e-04  1.045077e-03  1.060014e-03  1.038017e-03
#> [10,]  1.413373e-04  9.633700e-04  1.100203e-03  1.079659e-03  1.149489e-03
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0016169181 0.0010076010 0.0011154971 0.0011643701 0.0014616795
#>  [2,] 0.0006642224 0.0007703434 0.0009555620 0.0010743000 0.0012447773
#>  [3,] 0.0007432909 0.0007912280 0.0009575976 0.0010371555 0.0011543883
#>  [4,] 0.0009183959 0.0008511990 0.0010243007 0.0011287016 0.0013346965
#>  [5,] 0.0003238851 0.0007312415 0.0009365686 0.0010183132 0.0010922280
#>  [6,] 0.0009381249 0.0008471757 0.0010107048 0.0011115844 0.0013227101
#>  [7,] 0.0004309954 0.0007713043 0.0009836480 0.0011161139 0.0012194212
#>  [8,] 0.0002159511 0.0005112168 0.0006445434 0.0007144118 0.0007783924
#>  [9,] 0.0007217831 0.0004744017 0.0005243935 0.0005436483 0.0007283501
#> [10,] 0.0010398931 0.0008827012 0.0010367399 0.0011114197 0.0012726668
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0025383375 2.626492e-04 0.0031997954 0.0036006286 0.0032850849
#>  [2,] 0.0013919093 6.346870e-05 0.0014395213 0.0014420980 0.0016375206
#>  [3,] 0.0013308869 1.076985e-04 0.0014624764 0.0015316389 0.0016009159
#>  [4,] 0.0017112353 1.128198e-04 0.0019181244 0.0020259606 0.0021188673
#>  [5,] 0.0008662163 1.291368e-05 0.0006684860 0.0005269222 0.0007969923
#>  [6,] 0.0017262252 1.179378e-04 0.0019575750 0.0020892419 0.0021643544
#>  [7,] 0.0010673853 1.609105e-05 0.0009354035 0.0008194081 0.0011251544
#>  [8,] 0.0006099340 1.598697e-05 0.0004907617 0.0004431827 0.0006953311
#>  [9,] 0.0011973305 1.020452e-04 0.0014533013 0.0016658864 0.0015513067
#> [10,] 0.0017192637 1.598195e-04 0.0020271741 0.0022072588 0.0021330020
#>               V23         V24          V25          V26          V27
#>  [1,] 0.002307825 0.002474009 0.0028898313 0.0033611599 0.0033294871
#>  [2,] 0.001799088 0.001711791 0.0017349212 0.0018104061 0.0016697887
#>  [3,] 0.001649368 0.001617057 0.0016454238 0.0017416541 0.0016497995
#>  [4,] 0.002017421 0.001989042 0.0021095711 0.0022900425 0.0021704324
#>  [5,] 0.001368750 0.001207864 0.0010130474 0.0009132238 0.0008299552
#>  [6,] 0.002021950 0.002012410 0.0021500903 0.0023429378 0.0022314935
#>  [7,] 0.001621899 0.001474960 0.0013597321 0.0012771233 0.0011252568
#>  [8,] 0.001190810 0.001087274 0.0009180171 0.0008291648 0.0007183225
#>  [9,] 0.001039074 0.001118266 0.0013496872 0.0015714730 0.0015645013
#> [10,] 0.001836150 0.001868367 0.0020307323 0.0022600633 0.0021691487
#>                V28          V29           V3         V30          V31
#>  [1,] 0.0033541413 0.0031219582 2.486647e-04 0.002572955 0.0015012685
#>  [2,] 0.0016153330 0.0017704646 6.348706e-05 0.001709747 0.0014157547
#>  [3,] 0.0016507882 0.0016740760 1.132167e-04 0.001618500 0.0013131798
#>  [4,] 0.0021333409 0.0022031552 1.098717e-04 0.002023742 0.0015214294
#>  [5,] 0.0007993380 0.0009160068 2.239756e-05 0.001060131 0.0011826989
#>  [6,] 0.0021942940 0.0022385057 1.143885e-04 0.002029697 0.0014980472
#>  [7,] 0.0010479773 0.0012543519 2.588133e-05 0.001343028 0.0013178331
#>  [8,] 0.0007254471 0.0008311451 1.979665e-05 0.000894614 0.0009374350
#>  [9,] 0.0015566836 0.0014460392 9.218587e-05 0.001160305 0.0006486256
#> [10,] 0.0021606412 0.0021247914 1.633223e-04 0.001916884 0.0013632395
#>                V32          V33           V34           V35           V36
#>  [1,] 0.0015833458 0.0012689429  3.292329e-04 -0.0010753284 -1.978914e-03
#>  [2,] 0.0011443828 0.0012919032  1.415500e-03  0.0013778068  1.255576e-03
#>  [3,] 0.0011516358 0.0011915257  1.192022e-03  0.0011212615  9.850975e-04
#>  [4,] 0.0013206611 0.0013645322  1.233588e-03  0.0008551380  5.355267e-04
#>  [5,] 0.0008636591 0.0010325165  1.444362e-03  0.0019927229  2.298436e-03
#>  [6,] 0.0013109583 0.0013289633  1.152829e-03  0.0007292754  3.870108e-04
#>  [7,] 0.0009484967 0.0011669630  1.585766e-03  0.0020057993  2.157806e-03
#>  [8,] 0.0007429682 0.0009416647  1.291074e-03  0.0017290760  1.966270e-03
#>  [9,] 0.0006372581 0.0004677255 -3.895767e-05 -0.0007431552 -1.160086e-03
#> [10,] 0.0012747415 0.0012277155  9.585413e-04  0.0004804093  7.727392e-05
#>                 V37          V38          V39            V4          V40
#>  [1,] -0.0013109903 3.263704e-04 0.0010959553  5.052760e-04 0.0006202248
#>  [2,]  0.0011106505 1.082553e-03 0.0009207453  8.344350e-05 0.0010139906
#>  [3,]  0.0009227708 1.025001e-03 0.0009644499  1.548917e-04 0.0009534824
#>  [4,]  0.0005729063 9.441286e-04 0.0010014349  1.894819e-04 0.0009604791
#>  [5,]  0.0019639174 1.325341e-03 0.0008999129 -5.384441e-05 0.0011755469
#>  [6,]  0.0004596189 8.855807e-04 0.0009798866  2.032019e-04 0.0009167811
#>  [7,]  0.0018066356 1.266170e-03 0.0008341714 -2.685046e-05 0.0010824842
#>  [8,]  0.0016091993 1.062095e-03 0.0007429748 -3.876387e-05 0.0009340842
#>  [9,] -0.0007594023 3.126721e-05 0.0004079649  2.164767e-04 0.0002510833
#> [10,]  0.0002210719 7.994825e-04 0.0009867530  2.762102e-04 0.0008156419
#>                V41          V42          V43           V44           V45
#>  [1,] 0.0012581758 0.0018710944 1.907471e-03  1.968326e-03  0.0022447246
#>  [2,] 0.0008984827 0.0006735112 6.089065e-04  4.337014e-04  0.0003780647
#>  [3,] 0.0009287345 0.0008151539 7.366557e-04  6.025266e-04  0.0006085040
#>  [4,] 0.0010235910 0.0009916882 9.531299e-04  8.275556e-04  0.0008445616
#>  [5,] 0.0007858835 0.0002789958 9.875113e-05 -1.128985e-04 -0.0002600086
#>  [6,] 0.0010076681 0.0010171034 9.870726e-04  8.735088e-04  0.0009000613
#>  [7,] 0.0007639313 0.0002961837 1.910244e-04 -2.588663e-05 -0.0001617388
#>  [8,] 0.0006325136 0.0002921339 1.571487e-04 -4.692864e-05 -0.0001972697
#>  [9,] 0.0005190639 0.0008045148 8.302195e-04  8.602007e-04  0.0009747887
#> [10,] 0.0009898772 0.0011095000 1.084496e-03  1.005445e-03  0.0010994343
#>                 V46           V47           V48           V49           V5
#>  [1,]  1.968301e-03  1.437135e-03  1.080030e-03  5.704917e-04 6.330808e-04
#>  [2,]  4.177538e-04  3.538534e-04  1.947454e-04  1.174697e-04 1.591159e-04
#>  [3,]  5.785523e-04  4.390351e-04  2.992997e-04  1.737300e-04 2.207439e-04
#>  [4,]  8.072034e-04  6.275707e-04  4.155002e-04  2.302881e-04 2.812401e-04
#>  [5,] -1.623556e-04 -6.007795e-05 -1.129583e-04 -3.140366e-05 9.557774e-06
#>  [6,]  8.466851e-04  6.554336e-04  4.408242e-04  2.427653e-04 2.953006e-04
#>  [7,] -5.182384e-05  2.101873e-05 -5.752627e-05 -9.246432e-06 3.337397e-05
#>  [8,] -1.486376e-04 -7.573138e-05 -8.899825e-05 -2.334380e-05 1.338198e-05
#>  [9,]  8.587676e-04  6.632011e-04  4.805039e-04  2.518128e-04 2.801885e-04
#> [10,]  1.007604e-03  7.456146e-04  5.434541e-04  2.961319e-04 3.498938e-04
#>                V50           V51           V52           V53          V54
#>  [1,] 1.617368e-04  2.085081e-04  1.429241e-04  9.846167e-05 6.652422e-05
#>  [2,] 8.012486e-05  3.772092e-05  2.762143e-05  2.449614e-05 2.592050e-05
#>  [3,] 7.975769e-05  5.570227e-05  4.251399e-05  3.207553e-05 3.368739e-05
#>  [4,] 1.025630e-04  8.074668e-05  5.665808e-05  4.343355e-05 3.698464e-05
#>  [5,] 3.282652e-05 -2.695994e-05 -1.386918e-05 -5.924672e-06 1.114550e-05
#>  [6,] 1.024596e-04  8.567590e-05  6.003134e-05  4.533471e-05 3.742838e-05
#>  [7,] 5.001481e-05 -1.558273e-05 -6.242640e-06  7.360325e-07 1.409040e-05
#>  [8,] 2.478571e-05 -1.893504e-05 -1.120390e-05  9.379084e-08 1.276321e-05
#>  [9,] 6.759604e-05  9.327785e-05  6.345758e-05  4.149963e-05 2.340951e-05
#> [10,] 1.077377e-04  1.054702e-04  7.457545e-05  5.343012e-05 4.562321e-05
#>                V55           V56           V57          V58           V59
#>  [1,] 3.387026e-05  6.746535e-05  6.078007e-06 5.834129e-05  5.803068e-05
#>  [2,] 2.432054e-05  1.679900e-05  1.965168e-05 1.806276e-05  1.558137e-05
#>  [3,] 2.632900e-05  2.230315e-05  2.081787e-05 2.357582e-05  2.065890e-05
#>  [4,] 2.733528e-05  3.000281e-05  1.756958e-05 2.852012e-05  2.670981e-05
#>  [5,] 1.907670e-05 -3.775540e-06  2.582743e-05 1.302854e-06 -1.479453e-06
#>  [6,] 2.653436e-05  3.128051e-05  1.663787e-05 2.930366e-05  2.771464e-05
#>  [7,] 2.279027e-05  8.158277e-07  2.464805e-05 5.979091e-06  2.241699e-06
#>  [8,] 1.439012e-05 -1.513494e-06  2.045265e-05 3.845656e-06  6.593390e-07
#>  [9,] 7.655548e-06  2.845572e-05 -1.280193e-06 2.169811e-05  2.319834e-05
#> [10,] 2.999532e-05  3.721862e-05  1.605306e-05 3.580797e-05  3.392311e-05
#>                 V6          V60           V7           V8            V9
#>  [1,] 0.0004755851 3.530349e-05 0.0005328665 0.0006952783  1.558853e-03
#>  [2,] 0.0002285056 1.944015e-05 0.0003488000 0.0003073828  2.818641e-04
#>  [3,] 0.0002578328 2.107864e-05 0.0003499086 0.0003700607  4.789313e-04
#>  [4,] 0.0002991864 2.403318e-05 0.0004086887 0.0004194144  6.098455e-04
#>  [5,] 0.0001723111 1.111935e-05 0.0002784021 0.0002197679 -7.338291e-05
#>  [6,] 0.0003039010 2.409500e-05 0.0004065614 0.0004281685  6.527519e-04
#>  [7,] 0.0001719052 1.409801e-05 0.0002926687 0.0002129051 -3.690917e-05
#>  [8,] 0.0001296664 9.432884e-06 0.0002132671 0.0001542809 -9.954309e-05
#>  [9,] 0.0002268954 1.438410e-05 0.0002505515 0.0003370342  7.293086e-04
#> [10,] 0.0003202641 2.693452e-05 0.0004074164 0.0004697936  8.154919e-04
#> 
#> $vW[[2]]
#>              [,1]         [,2]        [,3]         [,4]        [,5]
#> [1,] -0.007108979  0.001550653 -0.01336874  0.002099543 -0.01057418
#> [2,]  0.007108979 -0.001550653  0.01336874 -0.002099543  0.01057418
#>              [,6]         [,7]        [,8]          [,9]       [,10]
#> [1,]  0.002331317 -0.001821707 -0.01630433  0.0002839128 -0.00880609
#> [2,] -0.002331317  0.001821707  0.01630433 -0.0002839128  0.00880609
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3324317 0.3914192 0.5882476 0.3111837 0.5581835 0.3309508 0.4087246
#>  [8] 0.7240536 0.2098562 0.4597558
#> 
#> $B[[2]]
#> [1] 0.10716301 0.01665195
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.003534500 0.002743535 0.002627198 0.003063074 0.002149764 0.003050074
#>  [7] 0.002452453 0.001786308 0.001524077 0.002932470
#> 
#> $vB[[2]]
#> [1] -0.01635152  0.01635152
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#>  [2,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#>  [3,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#>  [4,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [5,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#>  [6,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#>  [7,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#>  [8,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#>  [9,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [10,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [11,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [12,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [13,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [14,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [15,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [16,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [17,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [18,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [19,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [20,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [21,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [22,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [23,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [24,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [25,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [26,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [27,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [28,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [29,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [30,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [31,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [32,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [33,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [34,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [35,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [36,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [37,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [38,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [39,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [40,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#>  [2,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#>  [3,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#>  [4,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [5,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#>  [6,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#>  [7,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#>  [8,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#>  [9,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [10,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [11,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [12,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [13,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [14,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [15,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [16,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [17,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [18,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [19,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [20,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [21,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [22,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [23,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [24,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [25,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [26,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [27,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [28,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [29,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [30,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [31,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [32,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [33,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [34,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [35,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [36,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [37,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [38,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [39,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [40,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#>  [2,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#>  [3,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#>  [4,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [5,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#>  [6,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#>  [7,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#>  [8,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#>  [9,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [10,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [11,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [12,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [13,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [14,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [15,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [16,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [17,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [18,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [19,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [20,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [21,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [22,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [23,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [24,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [25,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [26,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [27,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [28,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [29,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [30,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [31,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [32,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [33,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [34,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [35,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [36,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [37,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [38,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [39,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [40,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#>  [2,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#>  [3,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#>  [4,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [5,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#>  [6,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#>  [7,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#>  [8,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#>  [9,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [10,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [11,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [12,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [13,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [14,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [15,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [16,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [17,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [18,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [19,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [20,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [21,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [22,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [23,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [24,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [25,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [26,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [27,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [28,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [29,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [30,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [31,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [32,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [33,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [34,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [35,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [36,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [37,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [38,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [39,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [40,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#>  [2,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#>  [3,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#>  [4,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [5,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#>  [6,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#>  [7,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#>  [8,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#>  [9,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [10,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [11,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [12,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [13,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [14,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [15,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [16,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [17,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [18,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [19,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [20,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [21,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [22,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [23,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [24,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [25,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [26,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [27,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [28,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [29,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [30,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [31,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [32,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [33,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [34,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [35,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [36,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [37,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [38,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [39,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [40,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#>  [2,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#>  [3,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#>  [4,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [5,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#>  [6,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#>  [7,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#>  [8,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#>  [9,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [10,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [11,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [12,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [13,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [14,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [15,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [16,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [17,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [18,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [19,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [20,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [21,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [22,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [23,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [24,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [25,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [26,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [27,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [28,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [29,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [30,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [31,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [32,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [33,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [34,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [35,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [36,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [37,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [38,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [39,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [40,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6685198 0.5996634 0.8547584 0.6427260 0.8566668 0.5670205 0.6186091
#>  [2,] 0.6842645 0.7004092 0.8407155 0.7243386 0.8795072 0.6188000 0.7308810
#>  [3,] 0.5078192 0.5115721 0.7406369 0.5023123 0.7332807 0.4817568 0.5966458
#>  [4,] 0.6230872 0.6879710 0.7934084 0.6869449 0.8450909 0.5958622 0.7477159
#>  [5,] 0.5137548 0.4882937 0.7604752 0.4993738 0.7551651 0.4482628 0.5907063
#>  [6,] 0.6351693 0.5379817 0.7962252 0.5923211 0.7864248 0.5603668 0.5687030
#>  [7,] 0.4663822 0.4462104 0.7008978 0.4538569 0.6869782 0.4384554 0.5877360
#>  [8,] 0.5034458 0.4767784 0.7367553 0.5069874 0.7308021 0.4788662 0.5680641
#>  [9,] 0.6310134 0.6572887 0.8053196 0.6274075 0.8272608 0.5475989 0.6816381
#> [10,] 0.4935957 0.5846739 0.6374375 0.5492170 0.7080140 0.5274670 0.7145049
#> [11,] 0.6327825 0.6664105 0.8050169 0.6364652 0.8327530 0.5548571 0.7049500
#> [12,] 0.5869646 0.5773845 0.8046592 0.5955950 0.8089671 0.5215227 0.6291569
#> [13,] 0.7176988 0.7160168 0.8517185 0.7159697 0.8841526 0.6236614 0.7020824
#> [14,] 0.6013405 0.5231412 0.7823774 0.6061429 0.7621128 0.5229252 0.5861379
#> [15,] 0.6682689 0.6332367 0.8542261 0.6609431 0.8627257 0.5791799 0.6631322
#> [16,] 0.5374013 0.6435304 0.7587620 0.6240778 0.8273215 0.5684614 0.7463592
#> [17,] 0.4412781 0.3516757 0.6759171 0.3828155 0.5783944 0.4220264 0.4609060
#> [18,] 0.5665401 0.5165644 0.7624134 0.5524712 0.7791217 0.5009619 0.5624707
#> [19,] 0.6162062 0.6179243 0.7697471 0.6253346 0.8267733 0.5544386 0.6767028
#> [20,] 0.4080606 0.5457690 0.5841028 0.4990851 0.6491109 0.4890416 0.7217464
#> [21,] 0.3915890 0.3261199 0.6756181 0.3700846 0.6385958 0.3945760 0.5219584
#> [22,] 0.4413236 0.5444228 0.6590570 0.4836380 0.6877421 0.4825805 0.6941334
#> [23,] 0.5484167 0.5640207 0.7501122 0.5739800 0.7735275 0.5120318 0.6946896
#> [24,] 0.6632220 0.6721789 0.8037480 0.6790495 0.8296610 0.5996740 0.6891579
#> [25,] 0.4515677 0.4650686 0.6540317 0.4583847 0.7097278 0.4109069 0.6127055
#> [26,] 0.6103687 0.5653871 0.7840213 0.5860750 0.7677954 0.5364814 0.6001600
#> [27,] 0.6389215 0.5971441 0.7870345 0.6311801 0.7990723 0.5799318 0.6386517
#> [28,] 0.4788721 0.5626898 0.6850774 0.4964639 0.7580333 0.4599474 0.6810768
#> [29,] 0.5625972 0.5997243 0.7546490 0.5577318 0.7833953 0.4976750 0.6835124
#> [30,] 0.5495530 0.4597590 0.7584703 0.5230533 0.7359231 0.5085617 0.5405634
#> [31,] 0.5660043 0.5398170 0.7745166 0.5476750 0.7778622 0.4957770 0.5839597
#> [32,] 0.5616235 0.5619072 0.7389910 0.5958269 0.7306813 0.5400542 0.6347317
#> [33,] 0.5579078 0.5051571 0.7517868 0.5680185 0.7431355 0.5212276 0.6110426
#> [34,] 0.6318959 0.5492192 0.8232978 0.6017795 0.8106086 0.5461863 0.6040171
#> [35,] 0.4880242 0.3805850 0.7418464 0.4262672 0.6756885 0.4444734 0.4782043
#> [36,] 0.4275481 0.5690468 0.6552534 0.5004897 0.7253272 0.4564495 0.7244709
#> [37,] 0.6701043 0.6512681 0.8424751 0.6592823 0.8605245 0.5609458 0.6711470
#> [38,] 0.4521641 0.4075591 0.7364037 0.4164028 0.6443742 0.4576136 0.5042571
#> [39,] 0.5486711 0.6110621 0.7421463 0.5763256 0.7565931 0.5090407 0.6583695
#> [40,] 0.6907314 0.6768749 0.8426209 0.6855086 0.8763920 0.5862435 0.6955039
#>            [,8]      [,9]     [,10]
#>  [1,] 0.9299693 0.3259874 0.8368396
#>  [2,] 0.9120364 0.3704643 0.8556552
#>  [3,] 0.8688279 0.2946545 0.7325335
#>  [4,] 0.8901817 0.3289523 0.8342891
#>  [5,] 0.8942270 0.2378727 0.7379657
#>  [6,] 0.8893558 0.3287862 0.7723368
#>  [7,] 0.8749963 0.1764535 0.7521154
#>  [8,] 0.8871387 0.2086001 0.7831503
#>  [9,] 0.8798728 0.3732859 0.7929142
#> [10,] 0.7844027 0.2666640 0.7230564
#> [11,] 0.8799810 0.3692347 0.7954208
#> [12,] 0.9046813 0.2979097 0.8029197
#> [13,] 0.8993953 0.4498241 0.8407473
#> [14,] 0.8890987 0.3236055 0.7630179
#> [15,] 0.9254902 0.3378842 0.8452780
#> [16,] 0.8877470 0.2420427 0.8330339
#> [17,] 0.8527370 0.2522960 0.6587723
#> [18,] 0.8835417 0.2454747 0.7790793
#> [19,] 0.8915774 0.2786154 0.8013750
#> [20,] 0.7870235 0.2270038 0.6862122
#> [21,] 0.8907486 0.1724379 0.6895855
#> [22,] 0.8006003 0.2625171 0.7101861
#> [23,] 0.8963107 0.2751904 0.7738963
#> [24,] 0.8660531 0.4011376 0.7920978
#> [25,] 0.8456339 0.1947735 0.7107378
#> [26,] 0.8835318 0.3301156 0.7641900
#> [27,] 0.8822896 0.3199362 0.7948611
#> [28,] 0.8479579 0.2152362 0.7508824
#> [29,] 0.8610151 0.3103157 0.7490753
#> [30,] 0.9000701 0.2415585 0.7545484
#> [31,] 0.8864751 0.2438452 0.7842517
#> [32,] 0.8736775 0.3102565 0.7677685
#> [33,] 0.8874055 0.3112187 0.7732313
#> [34,] 0.9154779 0.3323805 0.8047685
#> [35,] 0.8846017 0.2260055 0.7125826
#> [36,] 0.8262753 0.1874015 0.7470274
#> [37,] 0.9077343 0.3721514 0.8243317
#> [38,] 0.8742305 0.2430258 0.7329729
#> [39,] 0.8626561 0.2724624 0.7845070
#> [40,] 0.9007382 0.3998747 0.8368652
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5090452 0.4909548
#>  [2,] 0.5068302 0.4931698
#>  [3,] 0.5156321 0.4843679
#>  [4,] 0.5101700 0.4898300
#>  [5,] 0.5185234 0.4814766
#>  [6,] 0.5084741 0.4915259
#>  [7,] 0.5201011 0.4798989
#>  [8,] 0.5172750 0.4827250
#>  [9,] 0.5090117 0.4909883
#> [10,] 0.5142779 0.4857221
#> [11,] 0.5092709 0.4907291
#> [12,] 0.5129475 0.4870525
#> [13,] 0.5029891 0.4970109
#> [14,] 0.5101054 0.4898946
#> [15,] 0.5088307 0.4911693
#> [16,] 0.5164635 0.4835365
#> [17,] 0.5167583 0.4832417
#> [18,] 0.5140271 0.4859729
#> [19,] 0.5122458 0.4877542
#> [20,] 0.5192672 0.4807328
#> [21,] 0.5243812 0.4756188
#> [22,] 0.5177590 0.4822410
#> [23,] 0.5158118 0.4841882
#> [24,] 0.5054998 0.4945002
#> [25,] 0.5212433 0.4787567
#> [26,] 0.5097101 0.4902899
#> [27,] 0.5084750 0.4915250
#> [28,] 0.5199537 0.4800463
#> [29,] 0.5140110 0.4859890
#> [30,] 0.5147351 0.4852649
#> [31,] 0.5144388 0.4855612
#> [32,] 0.5112501 0.4887499
#> [33,] 0.5122922 0.4877078
#> [34,] 0.5099035 0.4900965
#> [35,] 0.5174443 0.4825557
#> [36,] 0.5219036 0.4780964
#> [37,] 0.5079518 0.4920482
#> [38,] 0.5172221 0.4827779
#> [39,] 0.5135828 0.4864172
#> [40,] 0.5060477 0.4939523
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]      [,3]         [,4]      [,5]         [,6]
#>  [1,]  0.70149791  0.40406269 1.7724206  0.587215562 1.7878773  0.269705019
#>  [2,]  0.77343949  0.84924710 1.6635613  0.966085734 1.9877720  0.484457848
#>  [3,]  0.03127940  0.04629680 1.0492814  0.009249193 1.0113319 -0.073005029
#>  [4,]  0.50267280  0.79065035 1.3455939  0.785874964 1.6966060  0.388253704
#>  [5,]  0.05503290 -0.04683366 1.1552868 -0.002504902 1.1263525 -0.207692345
#>  [6,]  0.55445808  0.15222024 1.3628667  0.373569017 1.3035082  0.242650787
#>  [7,] -0.13467449 -0.21599429 0.8515769 -0.185099083 0.7860296 -0.247433106
#>  [8,]  0.01378325 -0.09295317 1.0291719  0.027951372 0.9986960 -0.084585463
#>  [9,]  0.53656683  0.65123484 1.4198802  0.521110718 1.5663373  0.190973902
#> [10,] -0.02561843  0.34199020 0.5642593  0.197507417 0.8857580  0.109978901
#> [11,]  0.54417225  0.69199457 1.4179503  0.560054788 1.6052656  0.220315405
#> [12,]  0.35143113  0.31204545 1.4156734  0.387144219 1.4433128  0.086144150
#> [13,]  0.93307542  0.92478845 1.7481434  0.924556856 2.0323559  0.505118280
#> [14,]  0.41105376  0.09263115 1.2795749  0.431127761 1.1642977  0.091765023
#> [15,]  0.70036606  0.54612746 1.7681388  0.667499767 1.8381160  0.319407616
#> [16,]  0.14988528  0.59072023 1.1459041  0.506892781 1.5667618  0.275576589
#> [17,] -0.23597657 -0.61168177 0.7350712 -0.477614783 0.3161858 -0.314460395
#> [18,]  0.26774863  0.06628182 1.1659565  0.210660277 1.2605553  0.003847728
#> [19,]  0.47347624  0.48074725 1.2068838  0.512253663 1.5629293  0.218621108
#> [20,] -0.37198847  0.18358992 0.3396387 -0.003659626 0.6151334 -0.043840638
#> [21,] -0.44063781 -0.72578706 0.7337066 -0.531853922 0.5692748 -0.428117164
#> [22,] -0.23579200  0.17816097 0.6590946 -0.065471237 0.7895844 -0.069706080
#> [23,]  0.19427569  0.25749609 1.0992108  0.298108456 1.2283380  0.048136426
#> [24,]  0.67768569  0.71805628 1.4098859  0.749407102 1.5832263  0.404106875
#> [25,] -0.19433852 -0.13995375 0.6368088 -0.166846996 0.8940626 -0.360217612
#> [26,]  0.44886236  0.26305502 1.2892563  0.347762928 1.1959044  0.146185244
#> [27,]  0.57068612  0.39357968 1.3071420  0.537282722 1.3805061  0.322493325
#> [28,] -0.08456188  0.25208574 0.7772048 -0.014144763 1.1419275 -0.160554319
#> [29,]  0.25170932  0.40431659 1.1235629  0.231961661 1.2855636 -0.009300171
#> [30,]  0.19886490 -0.16131302 1.1443110  0.092278679 1.0248854  0.034250035
#> [31,]  0.26556688  0.15960600 1.2339928  0.191281164 1.2532518 -0.016892258
#> [32,]  0.24775355  0.24890613 1.0407310  0.388106966 0.9980821  0.160560830
#> [33,]  0.23267530  0.02062914 1.1081649  0.273771392 1.0623297  0.084961289
#> [34,]  0.54035893  0.19751637 1.5388522  0.412885314 1.4539698  0.185273340
#> [35,] -0.04791224 -0.48706591 1.0555873 -0.297097584 0.7340279 -0.223026094
#> [36,] -0.29186175  0.27796315 0.6422123  0.001958864 0.9710423 -0.174644549
#> [37,]  0.70865668  0.62461801 1.6767609  0.660097576 1.8196529  0.245001371
#> [38,] -0.19193046 -0.37406530 1.0273596 -0.337558020 0.5944006 -0.169953612
#> [39,]  0.19530276  0.45177878 1.0571538  0.307707552 1.1340911  0.036166926
#> [40,]  0.80354082  0.73944699 1.6778596  0.779204098 1.9586980  0.348457767
#>              [,7]     [,8]       [,9]     [,10]
#>  [1,]  0.48364867 2.586219 -0.7263900 1.6348987
#>  [2,]  0.99909679 2.338756 -0.5302255 1.7796628
#>  [3,]  0.39150868 1.890635 -0.8728843 1.0075147
#>  [4,]  1.08646750 2.092599 -0.7129274 1.6163355
#>  [5,]  0.36688610 2.134665 -1.1643780 1.0354218
#>  [6,]  0.27656127 2.084177 -0.7136800 1.2215535
#>  [7,]  0.35461392 1.945876 -1.5405624 1.1099262
#>  [8,]  0.27395710 2.061842 -1.3333842 1.2841200
#>  [9,]  0.76130998 1.991226 -0.5181458 1.3425816
#> [10,]  0.91736485 1.291510 -1.0116146 0.9596734
#> [11,]  0.87098201 1.992250 -0.5355015 1.3579158
#> [12,]  0.52860178 2.250357 -0.8572718 1.4046432
#> [13,]  0.85723365 2.190524 -0.2013815 1.6637985
#> [14,]  0.34802236 2.081568 -0.7372514 1.1692967
#> [15,]  0.67728362 2.519394 -0.6727371 1.6980359
#> [16,]  1.07928797 2.067931 -1.1415129 1.6072835
#> [17,] -0.15669595 1.756231 -1.0864041 0.6578280
#> [18,]  0.25119527 2.026405 -1.1228950 1.2603090
#> [19,]  0.73866028 2.106955 -0.9513400 1.3949104
#> [20,]  0.95314106 1.307076 -1.2253073 0.7824699
#> [21,]  0.08788994 2.098411 -1.5684470 0.7981823
#> [22,]  0.81951517 1.390050 -1.0329265 0.8962881
#> [23,]  0.82213627 2.156888 -0.9684458 1.2304444
#> [24,]  0.79618517 1.866503 -0.4007273 1.3376169
#> [25,]  0.45869897 1.700760 -1.4192865 0.8989700
#> [26,]  0.40613176 2.026309 -0.7076624 1.1757902
#> [27,]  0.56951711 2.014293 -0.7540652 1.3544804
#> [28,]  0.75872453 1.718673 -1.2936466 1.1033241
#> [29,]  0.76996063 1.823747 -0.7986440 1.0936866
#> [30,]  0.16261083 2.198004 -1.1441537 1.1230197
#> [31,]  0.33904975 2.055231 -1.1317127 1.2906175
#> [32,]  0.55257011 1.933873 -0.7989204 1.1957538
#> [33,]  0.45169677 2.064509 -0.7944279 1.2266479
#> [34,]  0.42223134 2.382433 -0.6974378 1.4163688
#> [35,] -0.08723806 2.036748 -1.2310054 0.9079601
#> [36,]  0.96674813 1.559456 -1.4669841 1.0828206
#> [37,]  0.71337740 2.286279 -0.5229983 1.5459755
#> [38,]  0.01702866 1.938893 -1.1361615 1.0097588
#> [39,]  0.65603673 1.837528 -0.9821650 1.2921271
#> [40,]  0.82597843 2.205454 -0.4059874 1.6350864
#> 
#> $pre[[3]]
#>                [,1]        [,2]
#>  [1,] -0.0122009794 -0.04838584
#>  [2,] -0.0243955408 -0.05171789
#>  [3,]  0.0110467729 -0.05150211
#>  [4,] -0.0156769454 -0.05636247
#>  [5,]  0.0181085377 -0.05601913
#>  [6,] -0.0070736358 -0.04097345
#>  [7,]  0.0219346242 -0.05851314
#>  [8,]  0.0135149733 -0.05561250
#>  [9,] -0.0130883840 -0.04913909
#> [10,]  0.0039854850 -0.05314154
#> [11,] -0.0133647280 -0.05045246
#> [12,] -0.0014806793 -0.05328211
#> [13,] -0.0319596641 -0.04391629
#> [14,] -0.0042900416 -0.04471719
#> [15,] -0.0150954773 -0.05042198
#> [16,]  0.0005884080 -0.06528957
#> [17,]  0.0244548071 -0.04260334
#> [18,]  0.0057368637 -0.05038634
#> [19,] -0.0056107900 -0.05460380
#> [20,]  0.0166802233 -0.06042685
#> [21,]  0.0386963502 -0.05890595
#> [22,]  0.0148462285 -0.05621962
#> [23,]  0.0042442868 -0.05902414
#> [24,] -0.0215640849 -0.04356411
#> [25,]  0.0244548209 -0.06056962
#> [26,] -0.0057234247 -0.04456857
#> [27,] -0.0116671107 -0.04557046
#> [28,]  0.0164008844 -0.06345617
#> [29,]  0.0023779740 -0.05368051
#> [30,]  0.0103717589 -0.04858576
#> [31,]  0.0057748103 -0.05199652
#> [32,] -0.0044665440 -0.04947444
#> [33,] -0.0003410507 -0.04951993
#> [34,] -0.0072253482 -0.04684448
#> [35,]  0.0232817691 -0.04652390
#> [36,]  0.0199732914 -0.06769703
#> [37,] -0.0169861936 -0.04879625
#> [38,]  0.0207389193 -0.04817655
#> [39,] -0.0006516238 -0.05499622
#> [40,] -0.0233589655 -0.04755107
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4909548 -0.4909548
#>  [2,] -0.5068302  0.5068302
#>  [3,] -0.5156321  0.5156321
#>  [4,] -0.5101700  0.5101700
#>  [5,]  0.4814766 -0.4814766
#>  [6,] -0.5084741  0.5084741
#>  [7,]  0.4798989 -0.4798989
#>  [8,] -0.5172750  0.5172750
#>  [9,]  0.4909883 -0.4909883
#> [10,] -0.5142779  0.5142779
#> [11,]  0.4907291 -0.4907291
#> [12,]  0.4870525 -0.4870525
#> [13,] -0.5029891  0.5029891
#> [14,] -0.5101054  0.5101054
#> [15,]  0.4911693 -0.4911693
#> [16,] -0.5164635  0.5164635
#> [17,]  0.4832417 -0.4832417
#> [18,] -0.5140271  0.5140271
#> [19,] -0.5122458  0.5122458
#> [20,] -0.5192672  0.5192672
#> [21,] -0.5243812  0.5243812
#> [22,]  0.4822410 -0.4822410
#> [23,]  0.4841882 -0.4841882
#> [24,] -0.5054998  0.5054998
#> [25,]  0.4787567 -0.4787567
#> [26,]  0.4902899 -0.4902899
#> [27,]  0.4915250 -0.4915250
#> [28,]  0.4800463 -0.4800463
#> [29,]  0.4859890 -0.4859890
#> [30,]  0.4852649 -0.4852649
#> [31,] -0.5144388  0.5144388
#> [32,] -0.5112501  0.5112501
#> [33,] -0.5122922  0.5122922
#> [34,]  0.4900965 -0.4900965
#> [35,]  0.4825557 -0.4825557
#> [36,]  0.4780964 -0.4780964
#> [37,]  0.4920482 -0.4920482
#> [38,] -0.5172221  0.5172221
#> [39,] -0.5135828  0.5135828
#> [40,]  0.4939523 -0.4939523
#> 
#> $L
#> [1] 0.6707321 0.7907348 1.0058471 0.8551490 0.6911509 0.6914036
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5217391