Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12          V13         V14
#>  [1,] -0.3174275 -0.12331115 -0.14915064  0.043795858 -0.014629465  0.01997919
#>  [2,] -0.3341698 -0.04095063 -0.02273989 -0.144972835 -0.114189830 -0.08214521
#>  [3,] -0.3361900 -0.10941645 -0.08852383 -0.179923369 -0.074691963 -0.11637709
#>  [4,] -0.2891996 -0.14988763  0.01034451 -0.066414679 -0.097141297  0.04945999
#>  [5,] -0.3621261 -0.09801722 -0.04224078 -0.007151726 -0.082611587 -0.07740841
#>  [6,] -0.3173425 -0.10931340 -0.08079720 -0.132811146 -0.132127879 -0.10638944
#>  [7,] -0.3574829 -0.21330614 -0.22254465 -0.083309677 -0.090064813 -0.05467590
#>  [8,] -0.2814445 -0.09242636 -0.04291139 -0.102531394  0.034367976 -0.10302612
#>  [9,] -0.2059791 -0.11103398 -0.14329913 -0.093235551 -0.008165552 -0.07683505
#> [10,] -0.2428337 -0.10048776 -0.05851850 -0.067892802 -0.115830879 -0.02287341
#>                V15           V16         V17          V18         V19
#>  [1,]  0.002762356 -0.0994805441 -0.03727546  0.007230512  0.01353974
#>  [2,] -0.030519655  0.0276315277 -0.01086104  0.055765538  0.06216808
#>  [3,] -0.072282736  0.0001440317  0.09379568  0.001101212  0.05011645
#>  [4,]  0.003937887  0.0545683889  0.08056124  0.060496626  0.13250596
#>  [5,] -0.067962237  0.0538767995  0.06134918  0.050948466 -0.03789804
#>  [6,] -0.007760436 -0.0070347984 -0.05878134 -0.015519243  0.06395629
#>  [7,] -0.151943118 -0.0952651453 -0.13802754 -0.212343116 -0.15417991
#>  [8,] -0.141313143  0.0423709095  0.03779754 -0.004438017  0.10330799
#>  [9,] -0.114950380 -0.0385876950 -0.05544270 -0.007242725 -0.13704612
#> [10,]  0.010820938 -0.1375191425 -0.02775670  0.098264773  0.11265065
#>               V2         V20          V21         V22          V23         V24
#>  [1,] -0.2696983  0.09699784  0.193516286  0.25675050  0.138435486  0.22244027
#>  [2,] -0.2044387  0.02998550  0.035109728  0.10735335  0.154468330  0.28226742
#>  [3,] -0.3061869  0.05550705  0.039674080  0.18583893  0.025280886  0.07703436
#>  [4,] -0.3541157  0.23877610  0.286567157  0.18992771  0.267926342  0.23814744
#>  [5,] -0.3128922  0.13085042  0.154564169  0.08607323  0.005634797  0.13146276
#>  [6,] -0.3405434 -0.03243399  0.066773909  0.12143085  0.237931643  0.12677686
#>  [7,] -0.2301627 -0.13012416 -0.092063142 -0.06400178 -0.010569576 -0.09351845
#>  [8,] -0.2761119  0.17349311  0.226618669  0.12296299  0.154214271  0.24856071
#>  [9,] -0.2762463 -0.06612452  0.082349781  0.06556252  0.057018415  0.10616698
#> [10,] -0.2256744  0.01342850  0.005920299  0.05256065  0.164348032  0.14395081
#>               V25         V26        V27         V28          V29         V3
#>  [1,]  0.22279291 0.286467095 0.30791717  0.25491068  0.199893110 -0.3072110
#>  [2,]  0.19440313 0.200419495 0.17424628  0.23160602  0.198847633 -0.2456833
#>  [3,]  0.07730694 0.004662639 0.11729832  0.04990920 -0.054555304 -0.2730267
#>  [4,]  0.19420292 0.115240756 0.18445493  0.06913500  0.210164974 -0.3611490
#>  [5,]  0.02213756 0.155302213 0.07668194  0.12091868 -0.005645478 -0.2859010
#>  [6,]  0.24746612 0.162604854 0.17746723  0.10856311  0.052518687 -0.3808561
#>  [7,] -0.02276777 0.047224499 0.05813606 -0.07789579  0.042296645 -0.1796851
#>  [8,]  0.21911311 0.262486210 0.23084323  0.24110563  0.171709825 -0.2622128
#>  [9,]  0.17147990 0.128517791 0.10153419  0.01536428  0.035490311 -0.3000452
#> [10,]  0.03400057 0.205216999 0.14939831  0.20904153  0.124206747 -0.1717571
#>               V30         V31          V32          V33          V34
#>  [1,]  0.16891909  0.03785335  0.132470142  0.059349932  0.037988143
#>  [2,]  0.09068832  0.12272259 -0.022879674  0.075830389 -0.027908527
#>  [3,]  0.03000362 -0.03362246 -0.112409554  0.002589456 -0.029334323
#>  [4,]  0.12918453  0.02909047  0.046101063  0.043760622  0.033611876
#>  [5,]  0.01692744  0.03586350 -0.053896025 -0.177840818  0.018400290
#>  [6,]  0.07013405  0.04916510  0.063616563  0.083929965 -0.079234656
#>  [7,] -0.01268323 -0.07171022 -0.222930598 -0.179888755 -0.096715435
#>  [8,]  0.19608594 -0.01068367  0.003803091 -0.023691448  0.003806573
#>  [9,] -0.06762812  0.00199439 -0.169195883 -0.143922378 -0.142323242
#> [10,]  0.02805567  0.05837310 -0.057030770 -0.052136477 -0.127167853
#>               V35         V36          V37         V38          V39         V4
#>  [1,] -0.11307431 -0.04484269  0.029861229 -0.09910840 -0.049346038 -0.3809100
#>  [2,]  0.03755010 -0.13114681 -0.055932038 -0.08090981  0.010239679 -0.1773341
#>  [3,] -0.08123133 -0.13078578 -0.140574877 -0.12283779 -0.190664934 -0.2361782
#>  [4,] -0.04119154  0.11340765  0.088392891 -0.06271718 -0.083952211 -0.3793179
#>  [5,] -0.02792464 -0.11384392 -0.149655991 -0.05578997 -0.069575815 -0.2522964
#>  [6,] -0.01844291 -0.03055930 -0.043008338  0.06837540  0.014785002 -0.2224410
#>  [7,] -0.12052612 -0.09758439 -0.161317425 -0.24239712 -0.245351861 -0.1989992
#>  [8,] -0.01921339  0.11275055 -0.005787805 -0.10303652  0.004215179 -0.1935432
#>  [9,] -0.05131039 -0.14507062 -0.032508708 -0.02435189 -0.089335464 -0.2456762
#> [10,] -0.09443359 -0.15263835 -0.048225751 -0.14722500 -0.143186919 -0.2290284
#>               V40          V41         V42          V43         V44         V45
#>  [1,] -0.13236326  0.001731656 -0.08912314 -0.027194480 -0.06294583 -0.04037613
#>  [2,]  0.00934100 -0.175207648 -0.17702541 -0.054489096 -0.14814300 -0.18057894
#>  [3,] -0.07418791 -0.087824387 -0.18423603 -0.194093950 -0.14364990 -0.15299237
#>  [4,]  0.07841800 -0.087570586  0.07626349 -0.093661853 -0.06963993 -0.06904745
#>  [5,] -0.08379401 -0.021794595 -0.13148988 -0.147188523 -0.07529039 -0.22187878
#>  [6,] -0.05627315 -0.010040440  0.04971977 -0.055414854 -0.14300870 -0.04165450
#>  [7,] -0.21091575 -0.104964760 -0.04698980 -0.185702075 -0.15067896 -0.25943485
#>  [8,] -0.04573784 -0.095825417 -0.12804070  0.003520592 -0.07251448 -0.07303501
#>  [9,] -0.13236870 -0.221333179 -0.16713335 -0.162069102 -0.20328634 -0.28026144
#> [10,] -0.18891312 -0.213846625 -0.17045133 -0.154273144 -0.14738915 -0.12675954
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.18252114 -0.16696891 -0.2094081 -0.2743900 -0.1754042 -0.3800595
#>  [2,] -0.03566138 -0.14614822 -0.2009680 -0.2652531 -0.2632949 -0.3342642
#>  [3,] -0.21930995 -0.20281425 -0.1898365 -0.3566668 -0.2186721 -0.3504272
#>  [4,] -0.21059227 -0.09455322 -0.2511561 -0.2569615 -0.2286732 -0.3764798
#>  [5,] -0.23748991 -0.27645103 -0.2219323 -0.2013578 -0.2764783 -0.3807126
#>  [6,] -0.17140284 -0.13770816 -0.2717635 -0.2888893 -0.3574382 -0.3349347
#>  [7,] -0.25412279 -0.25406268 -0.2834044 -0.1501844 -0.1658838 -0.2831326
#>  [8,] -0.15034798 -0.26350245 -0.2659018 -0.2349922 -0.3221596 -0.2244305
#>  [9,] -0.07929273 -0.26308973 -0.1189614 -0.2562781 -0.2155864 -0.1846999
#> [10,] -0.10337516 -0.12061665 -0.1916104 -0.2342382 -0.1444216 -0.3161557
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.4212236 -0.2947647 -0.4131662 -0.4208051 -0.4016625 -0.4843654
#>  [2,] -0.2536612 -0.3122464 -0.3920129 -0.2952519 -0.3573218 -0.3361361
#>  [3,] -0.2886281 -0.2516025 -0.2595191 -0.2944987 -0.3421520 -0.3117312
#>  [4,] -0.3097265 -0.3686470 -0.3604095 -0.2965106 -0.3461669 -0.3284472
#>  [5,] -0.3573014 -0.4295297 -0.3346858 -0.3543495 -0.2485449 -0.3257619
#>  [6,] -0.2684114 -0.2992203 -0.4571973 -0.4374241 -0.4526901 -0.3356065
#>  [7,] -0.2454742 -0.3023072 -0.2657826 -0.3199216 -0.2498295 -0.2939095
#>  [8,] -0.3114698 -0.3056803 -0.3306475 -0.3390050 -0.3847329 -0.3087557
#>  [9,] -0.3031400 -0.3355277 -0.3428336 -0.3398696 -0.2153347 -0.2645949
#> [10,] -0.3355008 -0.3370619 -0.2026984 -0.1843506 -0.3232825 -0.3297447
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.3995423 -0.3981222 -0.4287682 -0.2568193 -0.3353626 -0.1309557
#>  [2,] -0.2550113 -0.3653845 -0.2406199 -0.2130475 -0.3415541 -0.1932682
#>  [3,] -0.3777334 -0.3241081 -0.3225319 -0.1123225 -0.2473308 -0.1443383
#>  [4,] -0.4373336 -0.3582501 -0.3324333 -0.1787288 -0.3809291 -0.1326250
#>  [5,] -0.2924734 -0.3227005 -0.4130497 -0.2316384 -0.3138514 -0.2403438
#>  [6,] -0.4112500 -0.3325411 -0.4126555 -0.1375628 -0.4309411 -0.1319827
#>  [7,] -0.1919512 -0.3414273 -0.2730101 -0.2096510 -0.3072702 -0.2137844
#>  [8,] -0.3267416 -0.2607657 -0.3163943 -0.1782172 -0.3448270 -0.2444532
#>  [9,] -0.3648684 -0.3468930 -0.3701001 -0.3090046 -0.3392094 -0.2628722
#> [10,] -0.2873791 -0.1959602 -0.2015721 -0.2259806 -0.3081382 -0.2540787
#>                V8          V9
#>  [1,] -0.15936687 -0.16935818
#>  [2,] -0.17333586 -0.15408790
#>  [3,] -0.08835296 -0.23298526
#>  [4,] -0.14212282 -0.05522279
#>  [5,] -0.17559904 -0.04265949
#>  [6,] -0.16175238 -0.09518931
#>  [7,] -0.28885242 -0.08386396
#>  [8,] -0.21339988 -0.18101723
#>  [9,] -0.23872288 -0.21386369
#> [10,] -0.22709021 -0.08209161
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.08730184 -0.07540508 -0.06876980 -0.07845084 -0.03768883 -0.08542160
#> [2,]  0.04695058  0.06049815 -0.01262461  0.04697888  0.12330234  0.01220827
#>             [,7]         [,8]       [,9]      [,10]
#> [1,] -0.02313420 -0.058877097 -0.1067488 -0.0175898
#> [2,]  0.05647027  0.006923962  0.1270834  0.1477153
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 7.691505e-05 5.572987e-04 0.0007071509 0.0007614131 0.0007951956
#>  [2,] 1.048523e-04 7.655820e-04 0.0009719633 0.0010381910 0.0010866972
#>  [3,] 4.413139e-05 4.289294e-04 0.0004082973 0.0003636780 0.0005658350
#>  [4,] 6.816563e-05 4.874511e-04 0.0005987537 0.0006375239 0.0006718900
#>  [5,] 1.577408e-04 1.063486e-03 0.0014791490 0.0015888612 0.0014899220
#>  [6,] 6.740590e-05 5.900996e-04 0.0006726577 0.0007083289 0.0008483700
#>  [7,] 1.225125e-05 8.883276e-05 0.0001488048 0.0001598609 0.0001259543
#>  [8,] 5.032768e-05 4.979998e-04 0.0005086280 0.0005412084 0.0007018151
#>  [9,] 1.625327e-04 1.048497e-03 0.0016956037 0.0019169729 0.0015910377
#> [10,] 1.570004e-04 1.046936e-03 0.0014920129 0.0016223327 0.0015122918
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0010277867 0.0011789104 0.0013990087 0.0013805209 0.0014471219
#>  [2,] 0.0014307028 0.0016324594 0.0019290050 0.0019116004 0.0019972146
#>  [3,] 0.0009901715 0.0012912927 0.0016448429 0.0017352952 0.0017215195
#>  [4,] 0.0008708666 0.0009860252 0.0011700364 0.0011558042 0.0012091942
#>  [5,] 0.0017875142 0.0018408227 0.0020715823 0.0020367914 0.0022231742
#>  [6,] 0.0012040057 0.0014749600 0.0018209868 0.0018610045 0.0018934061
#>  [7,] 0.0001586434 0.0001538584 0.0001636133 0.0001476531 0.0001711243
#>  [8,] 0.0010123268 0.0012738938 0.0015980128 0.0016569320 0.0016806779
#>  [9,] 0.0016583269 0.0014858369 0.0015005459 0.0013817650 0.0016974430
#> [10,] 0.0017852213 0.0018127472 0.0020218798 0.0019870079 0.0021906870
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0015852782 1.213042e-04 0.0017208988 0.0016685692 0.0015897112
#>  [2,] 0.0022001341 1.622479e-04 0.0024160647 0.0023735008 0.0022849109
#>  [3,] 0.0016609540 6.600271e-05 0.0016248469 0.0017487198 0.0019874192
#>  [4,] 0.0013355289 1.061239e-04 0.0014652246 0.0014479774 0.0014150330
#>  [5,] 0.0027256920 2.258595e-04 0.0032435143 0.0031311282 0.0027659622
#>  [6,] 0.0019253799 1.112899e-04 0.0019269361 0.0019575662 0.0020622329
#>  [7,] 0.0002553548 1.846618e-05 0.0002981641 0.0002502665 0.0001755725
#>  [8,] 0.0016342293 9.086560e-05 0.0015339762 0.0015966574 0.0017944679
#>  [9,] 0.0025103955 2.329545e-04 0.0031619193 0.0028331243 0.0021073023
#> [10,] 0.0027269992 2.246557e-04 0.0032367347 0.0031051262 0.0027211391
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0016928929 0.0018441524 0.0018917664 0.0019629881 0.0019120176
#>  [2,] 0.0024407228 0.0026571080 0.0027336520 0.0028686435 0.0028149294
#>  [3,] 0.0022771907 0.0025453181 0.0028174583 0.0030292631 0.0029653450
#>  [4,] 0.0015528205 0.0017184932 0.0017887346 0.0018984825 0.0018849637
#>  [5,] 0.0028374323 0.0030054634 0.0028686614 0.0029832324 0.0029576267
#>  [6,] 0.0022900817 0.0025405298 0.0027354536 0.0028928962 0.0028289597
#>  [7,] 0.0001829349 0.0001977656 0.0001447582 0.0001482134 0.0001613375
#>  [8,] 0.0020463964 0.0022795151 0.0025011412 0.0026640988 0.0026131030
#>  [9,] 0.0020030183 0.0019990658 0.0014668514 0.0014109633 0.0014368050
#> [10,] 0.0027850165 0.0029337156 0.0027389142 0.0028195089 0.0027985254
#>                V28          V29           V3          V30          V31
#>  [1,] 0.0018612610 1.717551e-03 1.328148e-04 1.645420e-03 1.608878e-03
#>  [2,] 0.0026964489 2.461373e-03 1.773726e-04 2.327606e-03 2.255107e-03
#>  [3,] 0.0028267878 2.688675e-03 8.319362e-05 2.601287e-03 2.537973e-03
#>  [4,] 0.0018499111 1.708609e-03 1.158301e-04 1.619591e-03 1.558045e-03
#>  [5,] 0.0027223062 2.297777e-03 2.377222e-04 1.996105e-03 1.823826e-03
#>  [6,] 0.0027463971 2.602525e-03 1.287847e-04 2.504085e-03 2.463270e-03
#>  [7,] 0.0001182646 6.900956e-05 1.695349e-05 3.661558e-05 3.909412e-05
#>  [8,] 0.0025823383 2.494992e-03 1.106365e-04 2.422158e-03 2.406783e-03
#>  [9,] 0.0011230848 6.469393e-04 2.311083e-04 2.385606e-04 9.729820e-05
#> [10,] 0.0025406985 2.123809e-03 2.370387e-04 1.814482e-03 1.656719e-03
#>                V32          V33          V34           V35           V36
#>  [1,] 1.528159e-03 1.308540e-03 0.0013230810  1.369835e-03  0.0012877581
#>  [2,] 2.130938e-03 1.826111e-03 0.0018498942  1.903658e-03  0.0017602934
#>  [3,] 2.071361e-03 1.983507e-03 0.0020773282  2.276258e-03  0.0022783594
#>  [4,] 1.462568e-03 1.297706e-03 0.0012976406  1.324609e-03  0.0012445675
#>  [5,] 2.007256e-03 1.570843e-03 0.0015903515  1.429665e-03  0.0009913542
#>  [6,] 2.125210e-03 1.931829e-03 0.0019717159  2.116983e-03  0.0021201392
#>  [7,] 9.657635e-05 7.051468e-05 0.0000836204  3.065706e-05 -0.0000421071
#>  [8,] 2.014484e-03 1.882507e-03 0.0019051983  2.092909e-03  0.0021895763
#>  [9,] 7.737895e-04 3.011817e-04 0.0002998825 -1.674433e-04 -0.0008211526
#> [10,] 1.886667e-03 1.445736e-03 0.0014545367  1.254324e-03  0.0008026555
#>                 V37          V38          V39           V4         V40
#>  [1,]  1.203952e-03 1.150830e-03 0.0011386388 1.275654e-04 0.001068272
#>  [2,]  1.633838e-03 1.552740e-03 0.0015250343 1.732048e-04 0.001439417
#>  [3,]  1.930897e-03 1.429539e-03 0.0012068970 7.223382e-05 0.001408416
#>  [4,]  1.140386e-03 1.074705e-03 0.0010585261 1.136224e-04 0.001010991
#>  [5,]  1.042185e-03 1.401453e-03 0.0015444658 2.748813e-04 0.001200464
#>  [6,]  1.888804e-03 1.551159e-03 0.0014297844 1.113143e-04 0.001537216
#>  [7,] -4.614674e-06 7.269969e-05 0.0001026645 2.530182e-05 0.000082381
#>  [8,]  1.945541e-03 1.544833e-03 0.0014078759 8.317002e-05 0.001564952
#>  [9,] -3.982662e-04 5.773126e-04 0.0009859738 3.011304e-04 0.000422554
#> [10,]  8.845452e-04 1.304547e-03 0.0014841455 2.755926e-04 0.001118105
#>                V41          V42          V43          V44          V45
#>  [1,] 9.405500e-04 0.0008617606 0.0007508805 0.0006161656 0.0005751139
#>  [2,] 1.276061e-03 0.0011678835 0.0010323054 0.0008660194 0.0008051405
#>  [3,] 1.130980e-03 0.0007831115 0.0006343716 0.0005207841 0.0003322464
#>  [4,] 8.900890e-04 0.0008172723 0.0007073936 0.0005829317 0.0005429405
#>  [5,] 1.189080e-03 0.0013627701 0.0013394126 0.0011717065 0.0012301994
#>  [6,] 1.284564e-03 0.0010106818 0.0008420060 0.0006924865 0.0005406874
#>  [7,] 8.369396e-05 0.0001133274 0.0001190809 0.0001047099 0.0001251871
#>  [8,] 1.277303e-03 0.0009550138 0.0007616576 0.0006149196 0.0004311952
#>  [9,] 6.157373e-04 0.0011342802 0.0012797702 0.0011365321 0.0013461957
#> [10,] 1.130948e-03 0.0013436659 0.0013377413 0.0011716016 0.0012356784
#>                V46          V47          V48          V49           V5
#>  [1,] 0.0004816997 3.763746e-04 0.0002614359 1.546409e-04 1.890422e-04
#>  [2,] 0.0006662099 5.197777e-04 0.0003574156 2.129702e-04 2.646178e-04
#>  [3,] 0.0003137052 3.006511e-04 0.0001844584 9.796022e-05 1.716755e-04
#>  [4,] 0.0004542810 3.500023e-04 0.0002412539 1.423116e-04 1.731053e-04
#>  [5,] 0.0009644286 6.754112e-04 0.0004870209 3.064205e-04 3.669793e-04
#>  [6,] 0.0004744500 4.269026e-04 0.0002789119 1.580753e-04 2.208459e-04
#>  [7,] 0.0001003725 7.261795e-05 0.0000496322 3.482969e-05 3.778151e-05
#>  [8,] 0.0003876729 3.810224e-04 0.0002459015 1.352859e-04 1.900884e-04
#>  [9,] 0.0010175299 6.758368e-04 0.0005034112 3.430760e-04 3.750662e-04
#> [10,] 0.0009643226 6.767216e-04 0.0004905360 3.113118e-04 3.679515e-04
#>                V50          V51          V52          V53          V54
#>  [1,] 6.006260e-05 4.314251e-05 4.505548e-05 3.319301e-05 2.967531e-05
#>  [2,] 8.133761e-05 6.027130e-05 6.064388e-05 4.402300e-05 4.006038e-05
#>  [3,] 5.585892e-05 2.805385e-05 2.443136e-05 3.043502e-05 2.713359e-05
#>  [4,] 5.514464e-05 3.866110e-05 3.892088e-05 2.919552e-05 2.610034e-05
#>  [5,] 9.751896e-05 9.129855e-05 8.969009e-05 5.101491e-05 5.027166e-05
#>  [6,] 6.889921e-05 4.195111e-05 3.977511e-05 3.695745e-05 3.246969e-05
#>  [7,] 6.524150e-06 9.205284e-06 8.749300e-06 2.975007e-06 3.125834e-06
#>  [8,] 6.238902e-05 3.386766e-05 2.935234e-05 3.340043e-05 2.846611e-05
#>  [9,] 7.884991e-05 1.035024e-04 1.002630e-04 4.087098e-05 4.348145e-05
#> [10,] 9.641019e-05 9.321045e-05 9.016198e-05 5.058714e-05 5.052669e-05
#>                V55          V56          V57          V58          V59
#>  [1,] 4.076273e-05 2.816505e-05 2.256444e-05 2.940664e-05 2.783254e-05
#>  [2,] 5.530783e-05 3.825758e-05 3.058090e-05 3.919790e-05 3.767178e-05
#>  [3,] 2.868309e-05 2.279625e-05 2.763251e-05 1.782530e-05 2.354263e-05
#>  [4,] 3.446115e-05 2.423055e-05 2.031142e-05 2.548349e-05 2.444490e-05
#>  [5,] 7.473235e-05 4.870297e-05 3.151238e-05 5.653298e-05 4.651880e-05
#>  [6,] 3.970374e-05 3.041452e-05 2.930425e-05 2.686831e-05 3.001825e-05
#>  [7,] 7.145223e-06 4.010578e-06 1.901118e-06 5.935582e-06 3.971755e-06
#>  [8,] 3.034232e-05 2.533609e-05 2.672995e-05 2.107596e-05 2.504050e-05
#>  [9,] 7.789330e-05 4.647968e-05 1.718752e-05 6.186221e-05 4.130211e-05
#> [10,] 7.542428e-05 4.871028e-05 3.007498e-05 5.709394e-05 4.632775e-05
#>                 V6          V60           V7           V8           V9
#>  [1,] 2.906254e-04 2.168413e-05 0.0003258105 4.252880e-04 5.217384e-04
#>  [2,] 4.014357e-04 2.892059e-05 0.0004470620 5.686785e-04 7.205523e-04
#>  [3,] 3.038396e-04 2.181828e-05 0.0003935862 3.711352e-04 4.044984e-04
#>  [4,] 2.674932e-04 1.943525e-05 0.0003005330 3.754056e-04 4.581113e-04
#>  [5,] 4.749658e-04 3.120077e-05 0.0004688024 6.836682e-04 9.997597e-04
#>  [6,] 3.726986e-04 2.569184e-05 0.0004476228 4.905746e-04 5.476257e-04
#>  [7,] 3.649164e-05 1.830516e-06 0.0000302759 5.861583e-05 8.676681e-05
#>  [8,] 3.389538e-04 2.370249e-05 0.0004237410 4.334150e-04 4.386039e-04
#>  [9,] 3.901274e-04 2.071895e-05 0.0003019234 6.021799e-04 9.688641e-04
#> [10,] 4.642649e-04 2.996851e-05 0.0004491576 6.637333e-04 9.777734e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.03378385  0.03503885  0.02761866  0.03249281  0.02708609  0.03611277
#> [2,] -0.03378385 -0.03503885 -0.02761866 -0.03249281 -0.02708609 -0.03611277
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01505819  0.03335927  0.02786402  0.02717954
#> [2,] -0.01505819 -0.03335927 -0.02786402 -0.02717954
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.8665825  0.5445251  0.2370025  0.8071668  0.4498552  0.7970888
#>  [7] -0.1927513  0.4747141  0.3359148  0.2613594
#> 
#> $B[[2]]
#> [1] 0.01370424 0.12294130
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.003043700 0.004249348 0.004165770 0.002843535 0.004184890 0.004245865
#>  [7] 0.000214810 0.003997110 0.001995119 0.003968874
#> 
#> $vB[[2]]
#> [1]  0.0360538 -0.0360538
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#>  [2,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>  [3,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#>  [4,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#>  [5,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#>  [6,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#>  [7,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#>  [8,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#>  [9,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [10,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [11,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [12,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [13,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [14,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [15,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [16,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [17,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [18,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [19,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [20,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [21,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [22,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [23,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [24,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [25,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [26,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [27,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [28,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [29,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [30,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [31,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [32,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [33,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [34,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [35,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [36,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#> [37,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [38,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [39,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [40,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#>  [2,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>  [3,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#>  [4,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#>  [5,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#>  [6,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#>  [7,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#>  [8,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#>  [9,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [10,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [11,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [12,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [13,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [14,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [15,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [16,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [17,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [18,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [19,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [20,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [21,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [22,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [23,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [24,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [25,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [26,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [27,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [28,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [29,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [30,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [31,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [32,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [33,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [34,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [35,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [36,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#> [37,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [38,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [39,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [40,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#>  [2,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>  [3,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#>  [4,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#>  [5,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#>  [6,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#>  [7,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#>  [8,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#>  [9,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [10,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [11,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [12,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [13,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [14,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [15,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [16,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [17,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [18,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [19,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [20,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [21,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [22,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [23,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [24,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [25,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [26,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [27,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [28,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [29,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [30,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [31,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [32,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [33,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [34,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [35,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [36,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#> [37,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [38,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [39,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [40,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#>  [2,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>  [3,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#>  [4,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#>  [5,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#>  [6,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#>  [7,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#>  [8,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#>  [9,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [10,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [11,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [12,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [13,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [14,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [15,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [16,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [17,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [18,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [19,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [20,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [21,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [22,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [23,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [24,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [25,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [26,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [27,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [28,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [29,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [30,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [31,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [32,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [33,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [34,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [35,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [36,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#> [37,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [38,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [39,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [40,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#>  [2,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>  [3,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#>  [4,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#>  [5,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#>  [6,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#>  [7,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#>  [8,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#>  [9,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [10,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [11,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [12,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [13,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [14,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [15,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [16,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [17,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [18,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [19,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [20,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [21,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [22,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [23,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [24,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [25,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [26,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [27,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [28,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [29,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [30,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [31,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [32,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [33,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [34,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [35,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [36,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#> [37,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [38,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [39,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [40,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#>  [2,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#>  [3,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#>  [4,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#>  [5,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#>  [6,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#>  [7,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#>  [8,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#>  [9,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [10,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [11,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [12,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [13,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [14,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [15,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [16,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [17,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [18,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [19,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [20,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [21,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [22,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [23,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [24,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [25,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [26,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [27,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [28,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [29,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [30,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [31,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [32,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [33,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [34,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [35,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [36,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#> [37,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [38,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [39,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [40,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]       [,7]
#>  [1,] 0.8562095 0.7159202 0.3148678 0.8494426 0.4616243 0.7514809 0.08279058
#>  [2,] 0.7883572 0.6878837 0.4088742 0.8017034 0.5278326 0.7046859 0.14830436
#>  [3,] 0.8719041 0.7711456 0.4978825 0.9115996 0.6322633 0.7758546 0.10585644
#>  [4,] 0.7795780 0.6306375 0.2493560 0.8200093 0.3885883 0.7068617 0.06170619
#>  [5,] 0.9078583 0.8295828 0.4653900 0.8916781 0.5973425 0.8217036 0.17391672
#>  [6,] 0.8367391 0.7303452 0.4029778 0.8822120 0.5276422 0.7482844 0.09510754
#>  [7,] 0.9064552 0.8210983 0.5044333 0.9015443 0.6366993 0.8101190 0.17085851
#>  [8,] 0.8046673 0.6883371 0.4068295 0.8514136 0.5525233 0.7116718 0.10073522
#>  [9,] 0.8528341 0.7444212 0.3855410 0.8906015 0.5149331 0.7747834 0.08681240
#> [10,] 0.8674087 0.7628335 0.4512369 0.9085144 0.5996562 0.7582488 0.08882897
#> [11,] 0.9317548 0.8613095 0.5879475 0.9187433 0.6897606 0.8604038 0.22931528
#> [12,] 0.8159103 0.6832856 0.4490079 0.8699163 0.5914210 0.6977090 0.09036361
#> [13,] 0.9099526 0.8249890 0.4931427 0.9099257 0.6262463 0.8278815 0.15672552
#> [14,] 0.9004620 0.8037182 0.4629708 0.8939288 0.5994135 0.8217007 0.13894267
#> [15,] 0.9267615 0.8474113 0.5235896 0.9109998 0.6625563 0.8457124 0.18572260
#> [16,] 0.9235272 0.8273996 0.4484542 0.9292395 0.6115890 0.8400733 0.11146445
#> [17,] 0.9157882 0.8266434 0.4946249 0.9175030 0.6333866 0.8280293 0.16522566
#> [18,] 0.8955196 0.8071445 0.5438808 0.9220933 0.6660673 0.7837492 0.12195134
#> [19,] 0.8324759 0.6896934 0.3121618 0.8448000 0.4704534 0.7391733 0.08491845
#> [20,] 0.9045987 0.8263619 0.4760010 0.8792664 0.5967163 0.8284620 0.18479610
#> [21,] 0.9231697 0.8431320 0.5344887 0.9134414 0.6693486 0.8248936 0.16850663
#> [22,] 0.8680891 0.7587986 0.5062121 0.8611500 0.6459666 0.7586822 0.19254641
#> [23,] 0.9151862 0.8292969 0.4730924 0.9115911 0.6119834 0.8420542 0.14570963
#> [24,] 0.9016922 0.7857498 0.4588525 0.9071313 0.5959911 0.8177084 0.14064777
#> [25,] 0.9253630 0.8421536 0.5425959 0.9277889 0.6806206 0.8329677 0.13519301
#> [26,] 0.8230078 0.7002863 0.3562654 0.8785590 0.4933685 0.7384381 0.06104422
#> [27,] 0.7731025 0.6311111 0.4218515 0.8682944 0.5663244 0.6872586 0.09374668
#> [28,] 0.8727852 0.7843248 0.5366431 0.9003787 0.6519864 0.7904603 0.15347446
#> [29,] 0.7387592 0.5693248 0.2178628 0.8042126 0.3890907 0.6354285 0.04634574
#> [30,] 0.8984370 0.8035463 0.4713339 0.8802430 0.6048313 0.7957735 0.17411708
#> [31,] 0.8440284 0.7365103 0.4195608 0.8612447 0.5752926 0.7360985 0.12332538
#> [32,] 0.8795200 0.7923634 0.5195733 0.8471133 0.6407908 0.8016997 0.27662770
#> [33,] 0.8001631 0.6917493 0.4699009 0.8782093 0.5972192 0.7135607 0.10029699
#> [34,] 0.8615654 0.7615295 0.4775326 0.9080292 0.6205941 0.7558286 0.10072537
#> [35,] 0.8770950 0.7627215 0.3244220 0.8514332 0.4732753 0.7791035 0.10038866
#> [36,] 0.8850922 0.7282602 0.3200157 0.8703001 0.4937616 0.7667014 0.07886814
#> [37,] 0.8616229 0.7475448 0.4652321 0.9049841 0.6204174 0.7355775 0.09181523
#> [38,] 0.8758540 0.7614428 0.4120007 0.8549117 0.5506599 0.7660348 0.15713751
#> [39,] 0.7673271 0.6398378 0.3017600 0.7962385 0.4427375 0.7020557 0.11324255
#> [40,] 0.8344660 0.6962219 0.3522912 0.8544721 0.4732819 0.7427992 0.10105520
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7615493 0.2835181 0.4433874
#>  [2,] 0.7003613 0.3670488 0.4819671
#>  [3,] 0.8243499 0.4295098 0.5841120
#>  [4,] 0.7050859 0.2274542 0.3163560
#>  [5,] 0.8457091 0.4537234 0.6252753
#>  [6,] 0.7821892 0.3763983 0.5076764
#>  [7,] 0.8469629 0.5031185 0.6316222
#>  [8,] 0.7165931 0.3440851 0.5038004
#>  [9,] 0.7875009 0.3636031 0.4957268
#> [10,] 0.8106366 0.3886952 0.5621193
#> [11,] 0.8862114 0.5896546 0.7202957
#> [12,] 0.7540500 0.3494533 0.4985670
#> [13,] 0.8628360 0.4572920 0.6164976
#> [14,] 0.8374224 0.4367476 0.5856802
#> [15,] 0.8792002 0.5401204 0.6786964
#> [16,] 0.8966823 0.4532200 0.5976101
#> [17,] 0.8783458 0.4773572 0.6506533
#> [18,] 0.8576597 0.4385277 0.6332579
#> [19,] 0.7690866 0.3133991 0.4224111
#> [20,] 0.8435563 0.4791468 0.6052673
#> [21,] 0.8783590 0.5226206 0.6889739
#> [22,] 0.8049737 0.5005731 0.5985973
#> [23,] 0.8698612 0.4699572 0.6026508
#> [24,] 0.8460602 0.4471817 0.5613336
#> [25,] 0.8814543 0.4959788 0.6745914
#> [26,] 0.7582885 0.2729070 0.4150568
#> [27,] 0.7144243 0.3290635 0.4400815
#> [28,] 0.8334503 0.4663683 0.6152294
#> [29,] 0.6724761 0.1886305 0.2893856
#> [30,] 0.8310838 0.4566976 0.6237758
#> [31,] 0.7717977 0.3922625 0.5457141
#> [32,] 0.8149694 0.5509653 0.6345337
#> [33,] 0.7463317 0.3822861 0.4982889
#> [34,] 0.8132137 0.4038140 0.5734819
#> [35,] 0.7965962 0.3211026 0.4943849
#> [36,] 0.7914879 0.3058569 0.4783205
#> [37,] 0.8057274 0.3793000 0.5642472
#> [38,] 0.7903310 0.3918741 0.5627031
#> [39,] 0.6921652 0.2839498 0.3636452
#> [40,] 0.7526824 0.2915220 0.4213682
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4098783 0.5901217
#>  [2,] 0.4052500 0.5947500
#>  [3,] 0.3930240 0.6069760
#>  [4,] 0.4211876 0.5788124
#>  [5,] 0.3895292 0.6104708
#>  [6,] 0.4019332 0.5980668
#>  [7,] 0.3863857 0.6136143
#>  [8,] 0.4045515 0.5954485
#>  [9,] 0.4024653 0.5975347
#> [10,] 0.3967386 0.6032614
#> [11,] 0.3768283 0.6231717
#> [12,] 0.4032990 0.5967010
#> [13,] 0.3888148 0.6111852
#> [14,] 0.3921434 0.6078566
#> [15,] 0.3816878 0.6183122
#> [16,] 0.3898031 0.6101969
#> [17,] 0.3865612 0.6134388
#> [18,] 0.3891527 0.6108473
#> [19,] 0.4098466 0.5901534
#> [20,] 0.3891052 0.6108948
#> [21,] 0.3823948 0.6176052
#> [22,] 0.3893349 0.6106651
#> [23,] 0.3888815 0.6111185
#> [24,] 0.3925105 0.6074895
#> [25,] 0.3837245 0.6162755
#> [26,] 0.4109654 0.5890346
#> [27,] 0.4079459 0.5920541
#> [28,] 0.3894215 0.6105785
#> [29,] 0.4261050 0.5738950
#> [30,] 0.3901238 0.6098762
#> [31,] 0.3988430 0.6011570
#> [32,] 0.3846404 0.6153596
#> [33,] 0.4015280 0.5984720
#> [34,] 0.3952475 0.6047525
#> [35,] 0.4051124 0.5948876
#> [36,] 0.4061451 0.5938549
#> [37,] 0.3970345 0.6029655
#> [38,] 0.3976508 0.6023492
#> [39,] 0.4158501 0.5841499
#> [40,] 0.4102386 0.5897614
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]      [,2]         [,3]     [,4]        [,5]      [,6]       [,7]
#>  [1,] 1.784157 0.9243134 -0.777458788 1.730236 -0.15380534 1.1065259 -2.4050216
#>  [2,] 1.315051 0.7902438 -0.368621218 1.396975  0.11144547 0.8697124 -1.7479626
#>  [3,] 1.917900 1.2147911 -0.008469952 2.333325  0.54193863 1.2416702 -2.1337826
#>  [4,] 1.263209 0.5349526 -1.102050099 1.516410 -0.45325019 0.8801905 -2.7216789
#>  [5,] 2.287761 1.5826730 -0.138661937 2.107998  0.39440424 1.5279326 -1.5581190
#>  [6,] 1.634162 0.9963748 -0.393072784 2.013546  0.11068145 1.0894835 -2.2528079
#>  [7,] 2.271101 1.5238062  0.017733496 2.214502  0.56106664 1.4507835 -1.5795551
#>  [8,] 1.415725 0.7923564 -0.377087535 1.745731  0.21087095 0.9035175 -2.1890820
#>  [9,] 1.757004 1.0690760 -0.466094793 2.096900  0.05975027 1.2355212 -2.3531919
#> [10,] 1.878239 1.1682776 -0.195674547 2.295629  0.40403298 1.1431023 -2.3280178
#> [11,] 2.613963 1.8262089  0.355486760 2.425393  0.79900014 1.8186477 -1.2121816
#> [12,] 1.488881 0.7689124 -0.204679898 1.900219  0.36984293 0.8364118 -2.3092033
#> [13,] 2.313057 1.5505211 -0.027430784 2.312728  0.51614668 1.5706869 -1.6827965
#> [14,] 2.202368 1.4096973 -0.148388684 2.131515  0.40302208 1.5279125 -1.8240997
#> [15,] 2.537975 1.7144403  0.094428464 2.325905  0.67470674 1.7013612 -1.4780469
#> [16,] 2.491265 1.5673088 -0.206918190 2.575066  0.45399659 1.6587737 -2.0758689
#> [17,] 2.386450 1.5620224 -0.021501387 2.408894  0.54677290 1.5717248 -1.6198493
#> [18,] 2.148405 1.4315618  0.175975860 2.471134  0.69045112 1.2876503 -1.9740799
#> [19,] 1.603277 0.7986862 -0.790032107 1.694385 -0.11832422 1.0416760 -2.3773218
#> [20,] 2.249399 1.5600598 -0.096069719 1.985502  0.39180157 1.5747661 -1.4841852
#> [21,] 2.486214 1.6817186  0.138174327 2.356397  0.70524045 1.5498609 -1.5962482
#> [22,] 1.884168 1.1461039  0.024849564 1.824875  0.60135656 1.1454685 -1.4335484
#> [23,] 2.378670 1.5806524 -0.107734523 2.333219  0.45565724 1.6735923 -1.7686553
#> [24,] 2.216170 1.2994940 -0.164962924 2.279100  0.38878883 1.5008980 -1.8099202
#> [25,] 2.517549 1.6743402  0.170797572 2.553211  0.75662530 1.6068079 -1.8558029
#> [26,] 1.536860 0.8486615 -0.591610572 1.978854 -0.02652769 1.0378663 -2.7331698
#> [27,] 1.225913 0.5369863 -0.315177363 1.885962  0.26687023 0.7873341 -2.2687226
#> [28,] 1.925812 1.2910497  0.146835666 2.201439  0.62778206 1.3277022 -1.7076062
#> [29,] 1.039529 0.2790970 -1.278164511 1.412834 -0.45113598 0.5555770 -3.0241719
#> [30,] 2.179977 1.4086080 -0.114790293 1.994734  0.42563702 1.3600851 -1.5567250
#> [31,] 1.688512 1.0279090 -0.324576629 1.825667  0.30347845 1.0257879 -1.9613097
#> [32,] 1.987893 1.3392306  0.078333065 1.712137  0.57879820 1.3969516 -0.9612515
#> [33,] 1.387314 0.8083102 -0.120541959 1.975581  0.39389189 0.9127408 -2.1939290
#> [34,] 1.828353 1.1610836 -0.089930290 2.289805  0.49207061 1.1299443 -2.1891908
#> [35,] 1.965204 1.1676584 -0.733523686 1.745887 -0.10700089 1.2604493 -2.1929136
#> [36,] 2.041562 0.9858133 -0.753699471 1.903614 -0.02495488 1.1897782 -2.4578259
#> [37,] 1.828835 1.0855603 -0.139296202 2.253873  0.49132013 1.0231075 -2.2916697
#> [38,] 1.953741 1.1606057 -0.355700819 1.773656  0.20333750 1.1860551 -1.6796826
#> [39,] 1.193280 0.5746602 -0.838931033 1.362948 -0.23005944 0.8571061 -2.0580395
#> [40,] 1.617615 0.8293708 -0.608983095 1.770116 -0.10697443 1.0605687 -2.1855547
#>            [,8]         [,9]        [,10]
#>  [1,] 1.1611924 -0.927077073 -0.227425538
#>  [2,] 0.8490190 -0.544898717 -0.072162880
#>  [3,] 1.5461013 -0.283851407  0.339676903
#>  [4,] 0.8716355 -1.222742527 -0.770569167
#>  [5,] 1.7013354 -0.185637716  0.512000321
#>  [6,] 1.2784699 -0.504864177  0.030708185
#>  [7,] 1.7109762  0.012474012  0.539182211
#>  [8,] 0.9276245 -0.645142162  0.015201961
#>  [9,] 1.3099268 -0.559759452 -0.017093254
#> [10,] 1.4541516 -0.452800017  0.249767761
#> [11,] 2.0526136  0.362537707  0.945928744
#> [12,] 1.1203302 -0.621443309 -0.005732125
#> [13,] 1.8390475 -0.171249442  0.474708701
#> [14,] 1.6391735 -0.254372287  0.346135834
#> [15,] 1.9848776  0.160827216  0.747787597
#> [16,] 2.1608931 -0.187668724  0.395516997
#> [17,] 1.9768582 -0.090633148  0.621912251
#> [18,] 1.7959863 -0.247139646  0.546218764
#> [19,] 1.2031607 -0.784275727 -0.312883437
#> [20,] 1.6849306 -0.083461041  0.427461338
#> [21,] 1.9769818  0.090544183  0.795326407
#> [22,] 1.4176751  0.002292343  0.399623929
#> [23,] 1.8997324 -0.120316175  0.416522470
#> [24,] 1.7040285 -0.212064287  0.246575998
#> [25,] 2.0062749 -0.016085202  0.729025536
#> [26,] 1.1433189 -0.979923446 -0.343099445
#> [27,] 0.9169701 -0.712423901 -0.240831496
#> [28,] 1.6102803 -0.134730338  0.469347898
#> [29,] 0.7194057 -1.458933194 -0.898369821
#> [30,] 1.5933277 -0.173644773  0.505605649
#> [31,] 1.2184900 -0.437811786  0.183368505
#> [32,] 1.4826295  0.204571558  0.551716504
#> [33,] 1.0791426 -0.479856226 -0.006844243
#> [34,] 1.4710287 -0.389598517  0.296071591
#> [35,] 1.3651545 -0.748709103 -0.022461464
#> [36,] 1.3339177 -0.819560875 -0.086772335
#> [37,] 1.4224829 -0.492520383  0.258417212
#> [38,] 1.3269216 -0.439441289  0.252139754
#> [39,] 0.8102612 -0.924953025 -0.559577418
#> [40,] 1.1129699 -0.888003518 -0.317159178
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.1571236 0.2073448
#>  [2,] -0.1576166 0.2260205
#>  [3,] -0.1855476 0.2490709
#>  [4,] -0.1377413 0.1801590
#>  [5,] -0.1920785 0.2572132
#>  [6,] -0.1696745 0.2277420
#>  [7,] -0.1974101 0.2651202
#>  [8,] -0.1593237 0.2272121
#>  [9,] -0.1711178 0.2240855
#> [10,] -0.1782001 0.2408730
#> [11,] -0.2151061 0.2879263
#> [12,] -0.1633771 0.2283606
#> [13,] -0.1952608 0.2570360
#> [14,] -0.1892110 0.2491003
#> [15,] -0.2062212 0.2761694
#> [16,] -0.1959371 0.2522028
#> [17,] -0.1978519 0.2639384
#> [18,] -0.1926074 0.2582676
#> [19,] -0.1567246 0.2078752
#> [20,] -0.1941972 0.2568779
#> [21,] -0.2038229 0.2755729
#> [22,] -0.1878403 0.2622684
#> [23,] -0.1967361 0.2552802
#> [24,] -0.1900464 0.2467251
#> [25,] -0.2032576 0.2705116
#> [26,] -0.1566265 0.2033495
#> [27,] -0.1549752 0.2174881
#> [28,] -0.1913613 0.2583834
#> [29,] -0.1237827 0.1739780
#> [30,] -0.1889206 0.2578711
#> [31,] -0.1707860 0.2395021
#> [32,] -0.1963426 0.2735553
#> [33,] -0.1674360 0.2316666
#> [34,] -0.1802427 0.2450642
#> [35,] -0.1656130 0.2185948
#> [36,] -0.1635886 0.2163358
#> [37,] -0.1759265 0.2419105
#> [38,] -0.1746609 0.2406023
#> [39,] -0.1429479 0.1968849
#> [40,] -0.1573952 0.2055838
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5901217 -0.5901217
#>  [2,]  0.5947500 -0.5947500
#>  [3,]  0.6069760 -0.6069760
#>  [4,] -0.4211876  0.4211876
#>  [5,] -0.3895292  0.3895292
#>  [6,] -0.4019332  0.4019332
#>  [7,] -0.3863857  0.3863857
#>  [8,] -0.4045515  0.4045515
#>  [9,] -0.4024653  0.4024653
#> [10,]  0.6032614 -0.6032614
#> [11,] -0.3768283  0.3768283
#> [12,] -0.4032990  0.4032990
#> [13,] -0.3888148  0.3888148
#> [14,]  0.6078566 -0.6078566
#> [15,] -0.3816878  0.3816878
#> [16,]  0.6101969 -0.6101969
#> [17,] -0.3865612  0.3865612
#> [18,] -0.3891527  0.3891527
#> [19,]  0.5901534 -0.5901534
#> [20,] -0.3891052  0.3891052
#> [21,]  0.6176052 -0.6176052
#> [22,] -0.3893349  0.3893349
#> [23,] -0.3888815  0.3888815
#> [24,] -0.3925105  0.3925105
#> [25,]  0.6162755 -0.6162755
#> [26,] -0.4109654  0.4109654
#> [27,]  0.5920541 -0.5920541
#> [28,] -0.3894215  0.3894215
#> [29,]  0.5738950 -0.5738950
#> [30,]  0.6098762 -0.6098762
#> [31,] -0.3988430  0.3988430
#> [32,] -0.3846404  0.3846404
#> [33,]  0.5984720 -0.5984720
#> [34,]  0.6047525 -0.6047525
#> [35,]  0.5948876 -0.5948876
#> [36,]  0.5938549 -0.5938549
#> [37,]  0.6029655 -0.6029655
#> [38,]  0.6023492 -0.6023492
#> [39,] -0.4158501  0.4158501
#> [40,] -0.4102386  0.4102386
#> 
#> $L
#> [1] 0.6924543 0.7128916 0.7382638 0.6972477 0.8838990 0.6901207
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.6086957