Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11         V12          V13         V14
#>  [1,] -0.3175224 -0.122005334 -0.21302810 -0.05545845 -0.009810589 -0.08201022
#>  [2,] -0.2383188 -0.183542182 -0.10426549 -0.16901512 -0.110389778 -0.05422453
#>  [3,] -0.2481708 -0.008266823 -0.09609309 -0.10115241 -0.101714970 -0.07738216
#>  [4,] -0.3016760 -0.158225880 -0.03900013 -0.18187627 -0.135328582 -0.03128759
#>  [5,] -0.3554992 -0.157191281 -0.03544271 -0.02156218  0.006577562 -0.03614204
#>  [6,] -0.2932314 -0.185297611 -0.08396850 -0.01810676 -0.001657644 -0.13254473
#>  [7,] -0.3892259 -0.214642423 -0.06934538 -0.06759642 -0.152504981 -0.02229645
#>  [8,] -0.3682785 -0.031385533 -0.08030747 -0.02623219 -0.063764067  0.04779217
#>  [9,] -0.3394462 -0.046992905 -0.05002612  0.02415260 -0.045647205 -0.04586319
#> [10,] -0.2280284 -0.071654366 -0.12724753 -0.18455408 -0.015835895 -0.07512369
#>               V15         V16         V17         V18          V19         V2
#>  [1,]  0.06537040 -0.11666597 -0.04538562  0.07168918  0.052783919 -0.3212154
#>  [2,]  0.01166189 -0.04390753  0.02039746 -0.02446728 -0.059969994 -0.1971590
#>  [3,] -0.09481279  0.06460298  0.10248308 -0.02257698  0.027247427 -0.2935360
#>  [4,] -0.15602965 -0.14679998 -0.09346090  0.04458292  0.002152305 -0.3224744
#>  [5,] -0.01122529 -0.05430014  0.04573464  0.10656905  0.172623980 -0.2856516
#>  [6,] -0.12725140 -0.01720748 -0.03726748 -0.07039493  0.114893448 -0.2321294
#>  [7,] -0.10619369 -0.01108571  0.05086869  0.17994188  0.190861082 -0.4105644
#>  [8,]  0.07170481  0.11105277  0.08553504  0.05137339  0.069226898 -0.2531513
#>  [9,] -0.07243303  0.09700830  0.10523183  0.14896872  0.152647082 -0.2593593
#> [10,] -0.12173752 -0.01525296 -0.04704992 -0.02612301 -0.015011194 -0.1815748
#>                V20          V21          V22          V23        V24        V25
#>  [1,] -0.002357315  0.003872047 -0.006760357  0.168131077 0.09127713 0.10100012
#>  [2,] -0.048906291 -0.010470939  0.112642916  0.065459684 0.15356592 0.20534265
#>  [3,]  0.153036280  0.153753728  0.064768114  0.007286576 0.03736943 0.16668276
#>  [4,]  0.001255562 -0.014104854  0.157809962  0.043337730 0.15963604 0.08942632
#>  [5,]  0.042525240  0.098338804  0.162681961  0.214846468 0.14638171 0.10751871
#>  [6,]  0.084197464  0.037127536  0.063969609  0.067808776 0.10874890 0.13636401
#>  [7,]  0.216698805  0.110371122  0.227034675  0.329902977 0.29940640 0.19504325
#>  [8,]  0.148423053  0.168911580  0.065032095 -0.026866924 0.04876358 0.09459121
#>  [9,]  0.065592658  0.145504742  0.155698136  0.188348376 0.15234406 0.09906882
#> [10,]  0.045800781  0.113102517  0.025909482  0.138852346 0.21475722 0.08050558
#>              V26        V27         V28         V29         V3          V30
#>  [1,] 0.20972814 0.21205364  0.03722951  0.15382371 -0.3292610  0.073544147
#>  [2,] 0.25524591 0.14086559  0.16332863  0.11515340 -0.3006092 -0.011607513
#>  [3,] 0.08450560 0.05237816  0.16568356  0.19192752 -0.1915208  0.086304756
#>  [4,] 0.14620652 0.15029397  0.18913407  0.10508519 -0.1912145  0.037338711
#>  [5,] 0.10633915 0.13384809  0.12500400  0.05604036 -0.2557794  0.067857363
#>  [6,] 0.16325915 0.20735195  0.15180058  0.12027157 -0.2166920  0.079090939
#>  [7,] 0.16198339 0.15132606  0.25012579  0.18035755 -0.4019232  0.027685795
#>  [8,] 0.03760297 0.14283850  0.09493074 -0.04052247 -0.2983919 -0.029723443
#>  [9,] 0.13186122 0.22453189  0.23422445  0.08792458 -0.2092154  0.174787369
#> [10,] 0.07904558 0.12255525 -0.03786081 -0.10330846 -0.2150574 -0.008954243
#>               V31          V32          V33          V34         V35
#>  [1,] -0.02195791 -0.168107073 -0.076080834 -0.015000408 -0.16965461
#>  [2,] -0.11111754 -0.160185326 -0.010094407 -0.135512117 -0.01228099
#>  [3,]  0.16948551  0.085473584  0.014578591  0.037549870  0.06941412
#>  [4,]  0.06906212  0.016227064 -0.150010241 -0.132539534 -0.06221548
#>  [5,]  0.06648487  0.075771095 -0.137864854 -0.142633472 -0.07910072
#>  [6,]  0.01480464 -0.002745840  0.001238642  0.015541440 -0.05882495
#>  [7,]  0.05772777  0.004857638  0.064898106 -0.055527028  0.06329791
#>  [8,]  0.09240767  0.043729844 -0.050530003  0.004668857  0.04095536
#>  [9,]  0.01786072  0.058617887  0.051933742 -0.044035466 -0.17090070
#> [10,] -0.05100155 -0.151985560 -0.105977968 -0.178430994 -0.07068384
#>                V36          V37          V38         V39         V4         V40
#>  [1,] -0.131997814  0.050550857 -0.056924941 -0.14658635 -0.2980787 -0.10622552
#>  [2,] -0.032846577 -0.078991618 -0.047246659 -0.09561659 -0.2794750 -0.05926969
#>  [3,] -0.034129546 -0.017966981 -0.067212764  0.09416248 -0.1646847 -0.07754686
#>  [4,]  0.005756025 -0.033741692  0.037556906 -0.02699908 -0.3098858 -0.09753042
#>  [5,] -0.053150585 -0.019892478 -0.009496915 -0.14743942 -0.3089043 -0.03733474
#>  [6,]  0.062395146 -0.008777441 -0.100160473 -0.11280125 -0.3525226 -0.03264909
#>  [7,]  0.008172074  0.001799691  0.042552066 -0.08812562 -0.2302478  0.03730375
#>  [8,]  0.048884444 -0.049440580 -0.038884496 -0.02568981 -0.2750779 -0.04761784
#>  [9,] -0.033256196 -0.012320336 -0.077301900 -0.03023866 -0.1794213 -0.04799458
#> [10,] -0.103744026 -0.171162719 -0.179341591 -0.10453414 -0.1832143 -0.07047741
#>                 V41          V42         V43         V44          V45
#>  [1,] -0.1072522709 -0.152978704 -0.07050700 -0.12926656 -0.177945050
#>  [2,] -0.0866791057 -0.098514695 -0.19494004 -0.19907374 -0.208367798
#>  [3,] -0.0983253623 -0.149556513 -0.03770898 -0.10637965 -0.181526407
#>  [4,]  0.0005971987 -0.068514866 -0.19636777 -0.07286689 -0.159943506
#>  [5,] -0.1308436815 -0.126872552 -0.14656131 -0.20361526 -0.063686267
#>  [6,] -0.0674104106 -0.197622276 -0.07608477 -0.17457649 -0.195179170
#>  [7,]  0.0407415428  0.034849129 -0.10704414 -0.04488655 -0.147226550
#>  [8,] -0.0073351899 -0.005202923 -0.10047591 -0.19959561 -0.140282066
#>  [9,] -0.1391035988 -0.049249178 -0.11630262 -0.13911424  0.008154809
#> [10,] -0.2211392596 -0.096227044 -0.19916113 -0.07511678 -0.131134733
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.23345457 -0.1477280 -0.1794583 -0.2323585 -0.3356036 -0.3418460
#>  [2,] -0.13253308 -0.2523524 -0.1957751 -0.2241225 -0.1481332 -0.2868450
#>  [3,] -0.01935363 -0.1009089 -0.2646095 -0.3331297 -0.2884508 -0.2196427
#>  [4,] -0.20963419 -0.2624850 -0.2047098 -0.2663515 -0.2900164 -0.3215920
#>  [5,] -0.21375342 -0.3236916 -0.2457756 -0.2006293 -0.2492240 -0.3965855
#>  [6,] -0.17950598 -0.1434968 -0.3062576 -0.2696415 -0.2213758 -0.3902748
#>  [7,] -0.19515906 -0.1213011 -0.1743868 -0.4047891 -0.1748025 -0.3003237
#>  [8,] -0.15277194 -0.1867241 -0.3063509 -0.3276793 -0.3326461 -0.3078413
#>  [9,] -0.15458870 -0.1253730 -0.1411802 -0.1773563 -0.2044400 -0.2088219
#> [10,] -0.14714509 -0.1649959 -0.1200267 -0.2437561 -0.1946907 -0.3689541
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3852142 -0.2242124 -0.2814801 -0.3720759 -0.4476296 -0.2643835
#>  [2,] -0.3889313 -0.3678114 -0.4074023 -0.3534798 -0.2724751 -0.3884908
#>  [3,] -0.2471550 -0.2417315 -0.2918445 -0.3157139 -0.3374967 -0.2948134
#>  [4,] -0.3206418 -0.3699387 -0.1811960 -0.1801190 -0.3620308 -0.2131740
#>  [5,] -0.2830132 -0.2884924 -0.2635176 -0.3907660 -0.2585671 -0.2812737
#>  [6,] -0.3081483 -0.4100820 -0.3730289 -0.3794841 -0.2612831 -0.4277297
#>  [7,] -0.4136572 -0.4289877 -0.4016059 -0.3700662 -0.3741675 -0.4412176
#>  [8,] -0.3601166 -0.2920058 -0.3922863 -0.3613927 -0.3645209 -0.3712047
#>  [9,] -0.2438391 -0.2262790 -0.3647166 -0.2601789 -0.3860796 -0.2752968
#> [10,] -0.2883823 -0.2854140 -0.2302707 -0.2766245 -0.3907151 -0.2846188
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.3016238 -0.2264905 -0.3884380 -0.1456645 -0.3435622 -0.11315512
#>  [2,] -0.2666317 -0.3455523 -0.3588655 -0.1820367 -0.4120490 -0.13155652
#>  [3,] -0.2809456 -0.3894751 -0.2410538 -0.2326940 -0.3457253 -0.06416612
#>  [4,] -0.2694916 -0.3443114 -0.3674897 -0.1092047 -0.1682473 -0.19143010
#>  [5,] -0.3938590 -0.4174688 -0.4181178 -0.1678800 -0.3463506 -0.27021923
#>  [6,] -0.2542536 -0.3741312 -0.3101950 -0.1613044 -0.3358178 -0.21635918
#>  [7,] -0.5024895 -0.3652109 -0.3134808 -0.2257803 -0.4565772 -0.18670218
#>  [8,] -0.2898582 -0.3643890 -0.4326348 -0.2409481 -0.3449829 -0.13846239
#>  [9,] -0.3690767 -0.2327254 -0.2422455 -0.2088338 -0.3338149 -0.22278995
#> [10,] -0.2523221 -0.3019113 -0.3134693 -0.1676116 -0.2904107 -0.27078204
#>                V8          V9
#>  [1,] -0.20502944 -0.22699670
#>  [2,] -0.16197333 -0.24141605
#>  [3,] -0.17036529 -0.00052586
#>  [4,] -0.08534172 -0.15970789
#>  [5,] -0.12780609 -0.08071791
#>  [6,] -0.19833844 -0.15355726
#>  [7,] -0.25371787 -0.21811299
#>  [8,] -0.22383322 -0.05533543
#>  [9,] -0.11239959 -0.16463351
#> [10,] -0.21838301 -0.25546399
#> 
#> $W[[2]]
#>             [,1]       [,2]          [,3]       [,4]          [,5]         [,6]
#> [1,] -0.07747186 -0.1111315 -0.1574136124 -0.1099757 -0.1579925010 -0.078306062
#> [2,]  0.12578853  0.1041198 -0.0001999846  0.1650116  0.0002470749  0.004246265
#>            [,7]        [,8]         [,9]       [,10]
#> [1,] -0.1587184 -0.03117394 -0.007715496 -0.08451313
#> [2,]  0.1199941  0.07139954  0.008653384 -0.02380644
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 1.214048e-04 0.0008421782 0.0009712233 0.0010716672 0.0011959913
#>  [2,] 1.005407e-04 0.0006955640 0.0008299078 0.0008967304 0.0009479441
#>  [3,] 1.446091e-04 0.0010273293 0.0011242074 0.0012984107 0.0015815547
#>  [4,] 1.130574e-04 0.0007424985 0.0009735114 0.0009466352 0.0007999124
#>  [5,] 1.398650e-04 0.0010044050 0.0011002636 0.0012902684 0.0015457066
#>  [6,] 1.219421e-04 0.0009243239 0.0009350825 0.0012515304 0.0017161344
#>  [7,] 1.130148e-04 0.0007028697 0.0008040235 0.0008287085 0.0008900720
#>  [8,] 1.191460e-04 0.0009204773 0.0009340260 0.0011950463 0.0016135445
#>  [9,] 1.191000e-04 0.0009731710 0.0010164791 0.0014101733 0.0018477300
#> [10,] 8.650719e-05 0.0006555487 0.0005971066 0.0007383590 0.0011403257
#>                V14          V15           V16           V17           V18
#>  [1,] 0.0012210734 1.413873e-03  1.692188e-03  0.0018536654  1.932570e-03
#>  [2,] 0.0008844053 9.326790e-04  1.106092e-03  0.0012044293  1.290917e-03
#>  [3,] 0.0018434875 2.383818e-03  2.873886e-03  0.0031628508  3.205166e-03
#>  [4,] 0.0003790150 6.477534e-06 -5.618389e-05 -0.0001696127 -9.137099e-05
#>  [5,] 0.0017652961 2.257532e-03  2.707679e-03  0.0029803367  3.047286e-03
#>  [6,] 0.0023258095 3.323336e-03  4.064475e-03  0.0046133481  4.674041e-03
#>  [7,] 0.0008507541 9.552115e-04  1.103226e-03  0.0011097724  1.101600e-03
#>  [8,] 0.0021325218 2.949069e-03  3.579539e-03  0.0040333177  4.104077e-03
#>  [9,] 0.0023628240 3.299054e-03  3.989865e-03  0.0044836408  4.556878e-03
#> [10,] 0.0017165471 2.427545e-03  2.982108e-03  0.0033859687  3.450677e-03
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0022718104 1.473783e-04 0.0025504371 0.0026346506 0.0025664213
#>  [2,] 0.0016077300 1.255782e-04 0.0018719088 0.0019823689 0.0019185597
#>  [3,] 0.0035077314 1.653433e-04 0.0037255753 0.0036988233 0.0036328919
#>  [4,] 0.0003590764 1.538802e-04 0.0007579336 0.0009795684 0.0008498159
#>  [5,] 0.0033615759 1.619365e-04 0.0035993297 0.0035765916 0.0035312695
#>  [6,] 0.0047745231 1.211641e-04 0.0047746678 0.0045738804 0.0045812661
#>  [7,] 0.0013539486 1.412212e-04 0.0016181634 0.0016597255 0.0014919512
#>  [8,] 0.0042881994 1.246567e-04 0.0043983746 0.0042754930 0.0043137528
#>  [9,] 0.0047193955 1.196292e-04 0.0047344497 0.0044534754 0.0045136810
#> [10,] 0.0035953008 8.926476e-05 0.0037311040 0.0037249372 0.0037594597
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0027038538 0.0028204303 0.0027162619 0.0030120476 0.0031371862
#>  [2,] 0.0019278453 0.0019855416 0.0019225361 0.0020985566 0.0021748397
#>  [3,] 0.0040779929 0.0043221806 0.0041289546 0.0046552632 0.0048803098
#>  [4,] 0.0003999421 0.0002938965 0.0003206936 0.0003228788 0.0003462327
#>  [5,] 0.0039987371 0.0042312190 0.0040594733 0.0045769505 0.0047981685
#>  [6,] 0.0055415315 0.0059548697 0.0056556332 0.0063313512 0.0066199903
#>  [7,] 0.0014482550 0.0015111758 0.0014587218 0.0016995265 0.0018074597
#>  [8,] 0.0051689833 0.0055126456 0.0052588585 0.0059166439 0.0061833727
#>  [9,] 0.0056309744 0.0060056172 0.0057020865 0.0064969617 0.0068903833
#> [10,] 0.0044755574 0.0048331833 0.0046476244 0.0051394056 0.0052784238
#>                V28         V29           V3         V30          V31
#>  [1,] 0.0029647926 0.002612519 0.0001727918 0.002156396  0.001845007
#>  [2,] 0.0020883948 0.001838888 0.0001347877 0.001504319  0.001199549
#>  [3,] 0.0044963462 0.003963872 0.0002325643 0.003299689  0.003055943
#>  [4,] 0.0005378315 0.000492549 0.0001161299 0.000349579 -0.000137953
#>  [5,] 0.0044758793 0.003991986 0.0002203207 0.003320028  0.003046038
#>  [6,] 0.0059580800 0.005262973 0.0002416172 0.004408122  0.004403187
#>  [7,] 0.0018033419 0.001634196 0.0001474468 0.001386640  0.001119309
#>  [8,] 0.0056395946 0.005032093 0.0002216288 0.004213717  0.004063236
#>  [9,] 0.0063235803 0.005721246 0.0002364237 0.004742243  0.004706367
#> [10,] 0.0047260176 0.004087517 0.0001695483 0.003446281  0.003372974
#>                V32           V33           V34           V35           V36
#>  [1,] 1.705699e-03  0.0017326609  0.0016951598  0.0016345429  0.0015622574
#>  [2,] 1.141663e-03  0.0011153760  0.0010173798  0.0008962792  0.0008226143
#>  [3,] 2.698518e-03  0.0028627258  0.0029890115  0.0031011083  0.0030778263
#>  [4,] 6.571859e-05 -0.0001438748 -0.0004913395 -0.0008356006 -0.0009270327
#>  [5,] 2.710944e-03  0.0029044507  0.0030498062  0.0031398303  0.0030854737
#>  [6,] 3.741874e-03  0.0041161956  0.0045204428  0.0048726210  0.0049273250
#>  [7,] 1.070802e-03  0.0010707844  0.0010114202  0.0009569026  0.0009146194
#>  [8,] 3.513236e-03  0.0038898953  0.0042459493  0.0045165373  0.0044826441
#>  [9,] 4.008885e-03  0.0044643880  0.0049833035  0.0053673968  0.0054088262
#> [10,] 2.922151e-03  0.0031593584  0.0034246014  0.0036708439  0.0035863482
#>                 V37          V38          V39           V4          V40
#>  [1,]  0.0014551748 0.0011192864 0.0010213182 0.0002022258 0.0011169740
#>  [2,]  0.0008286110 0.0007540470 0.0007669595 0.0001712121 0.0007385353
#>  [3,]  0.0027195461 0.0017720190 0.0014146007 0.0002287476 0.0018508883
#>  [4,] -0.0005722042 0.0001203191 0.0004751986 0.0001928734 0.0000201022
#>  [5,]  0.0026936537 0.0017958418 0.0014607621 0.0002220774 0.0018667635
#>  [6,]  0.0042070928 0.0023858697 0.0016505183 0.0001765909 0.0026085197
#>  [7,]  0.0008929146 0.0008034303 0.0007768527 0.0001745742 0.0007336893
#>  [8,]  0.0038005353 0.0022993383 0.0016676383 0.0001886361 0.0024105510
#>  [9,]  0.0044944748 0.0026503802 0.0019438415 0.0001655477 0.0029521658
#> [10,]  0.0030485350 0.0017887222 0.0012374497 0.0001572524 0.0017866082
#>                V41          V42          V43          V44          V45
#>  [1,] 1.006694e-03 0.0009742638 0.0008737231 0.0007444291 6.070621e-04
#>  [2,] 6.788162e-04 0.0007214208 0.0006945224 0.0006328532 5.736564e-04
#>  [3,] 1.632439e-03 0.0014111117 0.0011376308 0.0008580736 5.270998e-04
#>  [4,] 9.799381e-05 0.0004230258 0.0005951596 0.0007073242 8.961819e-04
#>  [5,] 1.632204e-03 0.0014227169 0.0011574105 0.0008580956 5.133927e-04
#>  [6,] 2.247127e-03 0.0017269779 0.0012189242 0.0007251083 9.130959e-05
#>  [7,] 6.545318e-04 0.0006871249 0.0006443485 0.0005805278 5.700956e-04
#>  [8,] 2.060748e-03 0.0016133033 0.0011884111 0.0007397752 1.805720e-04
#>  [9,] 2.511868e-03 0.0019825449 0.0014391191 0.0008383851 5.900113e-05
#> [10,] 1.518589e-03 0.0011205579 0.0008186586 0.0005268687 1.517733e-04
#>                 V46          V47          V48          V49           V5
#>  [1,]  4.520206e-04 0.0003683102 0.0003007408 0.0001789976 0.0002588928
#>  [2,]  4.311310e-04 0.0003172339 0.0002512648 0.0001513219 0.0002112012
#>  [3,]  3.768305e-04 0.0004133338 0.0003563963 0.0002076086 0.0003192310
#>  [4,]  7.022592e-04 0.0003767487 0.0002772186 0.0001751377 0.0001886676
#>  [5,]  3.653053e-04 0.0003990006 0.0003437100 0.0001984165 0.0003201287
#>  [6,]  1.247017e-05 0.0003150959 0.0003021222 0.0001659254 0.0003159264
#>  [7,]  4.419093e-04 0.0003067235 0.0002501121 0.0001498992 0.0001962408
#>  [8,]  8.218917e-05 0.0003045992 0.0002945368 0.0001614676 0.0003234265
#>  [9,] -3.186878e-05 0.0003048763 0.0003007797 0.0001605969 0.0003407525
#> [10,]  7.667041e-05 0.0002394919 0.0002300109 0.0001261728 0.0002599214
#>                V50          V51          V52          V53          V54
#>  [1,] 7.300139e-05 5.925037e-05 4.471782e-05 3.967847e-05 4.392312e-05
#>  [2,] 5.410018e-05 4.854437e-05 3.811378e-05 2.843270e-05 3.268297e-05
#>  [3,] 1.039320e-04 7.261832e-05 5.044754e-05 5.881551e-05 6.146031e-05
#>  [4,] 3.262175e-05 4.826183e-05 4.618267e-05 1.244367e-05 1.992940e-05
#>  [5,] 1.006400e-04 7.161354e-05 4.945263e-05 5.911027e-05 6.056618e-05
#>  [6,] 1.205477e-04 6.744528e-05 3.503730e-05 7.285648e-05 7.078854e-05
#>  [7,] 5.585867e-05 4.749845e-05 4.061248e-05 2.878127e-05 3.153735e-05
#>  [8,] 1.120400e-04 6.778047e-05 3.961769e-05 7.095161e-05 6.921893e-05
#>  [9,] 1.214853e-04 7.318636e-05 3.824976e-05 8.235405e-05 7.522395e-05
#> [10,] 9.218305e-05 5.119765e-05 3.014348e-05 5.477247e-05 5.550021e-05
#>                V55          V56           V57          V58          V59
#>  [1,] 2.897280e-05 3.253166e-05  3.072493e-05 2.543222e-05 2.281420e-05
#>  [2,] 2.280058e-05 2.359756e-05  1.911246e-05 2.066246e-05 1.880478e-05
#>  [3,] 3.648761e-05 4.764730e-05  5.361591e-05 3.068000e-05 2.649709e-05
#>  [4,] 1.817097e-05 1.096416e-05 -5.347091e-06 2.102601e-05 2.009822e-05
#>  [5,] 3.666591e-05 4.714126e-05  5.245845e-05 3.045287e-05 2.643308e-05
#>  [6,] 3.667514e-05 5.841177e-05  7.783797e-05 2.756916e-05 2.089553e-05
#>  [7,] 2.360025e-05 2.299544e-05  1.954009e-05 2.242086e-05 1.931678e-05
#>  [8,] 3.802365e-05 5.526172e-05  6.947606e-05 2.849617e-05 2.351298e-05
#>  [9,] 3.790542e-05 6.184644e-05  8.069017e-05 2.929167e-05 2.277740e-05
#> [10,] 3.336810e-05 4.505331e-05  5.862054e-05 2.111898e-05 1.904307e-05
#>                 V6          V60           V7           V8           V9
#>  [1,] 0.0003921301 1.847166e-05 0.0004607146 0.0005002238 0.0007254984
#>  [2,] 0.0002907796 1.545189e-05 0.0003410598 0.0003921285 0.0005966020
#>  [3,] 0.0005660478 2.042070e-05 0.0006677175 0.0006696401 0.0008857236
#>  [4,] 0.0001418094 1.698921e-05 0.0001594735 0.0003226182 0.0006168848
#>  [5,] 0.0005617133 2.082357e-05 0.0006571129 0.0006557525 0.0008635164
#>  [6,] 0.0007070092 1.580379e-05 0.0008347283 0.0007284455 0.0008272724
#>  [7,] 0.0002692535 1.556530e-05 0.0003249221 0.0004093561 0.0006175659
#>  [8,] 0.0006572082 1.767350e-05 0.0007790487 0.0006845697 0.0008122717
#>  [9,] 0.0007470903 1.587936e-05 0.0008539555 0.0007586696 0.0008276169
#> [10,] 0.0005302825 1.518787e-05 0.0006419659 0.0005133537 0.0006091765
#> 
#> $vW[[2]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]       [,6]       [,7]
#> [1,]  0.123445  0.1158275  0.142759  0.1144554  0.1348831  0.1417177  0.1657549
#> [2,] -0.123445 -0.1158275 -0.142759 -0.1144554 -0.1348831 -0.1417177 -0.1657549
#>            [,8]       [,9]       [,10]
#> [1,]  0.1437943  0.1508648  0.08914491
#> [2,] -0.1437943 -0.1508648 -0.08914491
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5205079 0.2976248 0.5255001 0.3006846 0.4625794 0.6322098 0.8784310
#>  [8] 0.6638138 0.5740627 0.2631424
#> 
#> $B[[2]]
#> [1] -0.11134401  0.07168107
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0040124888 0.0027499761 0.0063451018 0.0003126884 0.0062506503
#>  [6] 0.0087280648 0.0024746232 0.0080882329 0.0091345550 0.0067407785
#> 
#> $vB[[2]]
#> [1]  0.1801363 -0.1801363
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>  [2,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#>  [3,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#>  [4,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#>  [5,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#>  [6,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#>  [7,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#>  [8,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [9,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [10,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [11,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [12,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [13,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [14,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [15,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [16,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [17,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [18,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [19,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [20,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [21,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [22,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [23,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [24,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [25,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [26,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [27,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [28,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [29,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [30,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [31,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [32,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [33,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [34,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [35,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [36,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [37,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [38,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [39,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [40,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>  [2,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#>  [3,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#>  [4,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#>  [5,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#>  [6,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#>  [7,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#>  [8,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [9,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [10,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [11,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [12,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [13,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [14,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [15,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [16,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [17,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [18,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [19,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [20,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [21,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [22,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [23,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [24,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [25,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [26,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [27,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [28,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [29,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [30,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [31,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [32,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [33,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [34,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [35,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [36,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [37,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [38,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [39,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [40,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>  [2,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#>  [3,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#>  [4,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#>  [5,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#>  [6,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#>  [7,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#>  [8,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [9,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [10,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [11,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [12,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [13,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [14,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [15,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [16,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [17,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [18,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [19,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [20,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [21,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [22,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [23,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [24,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [25,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [26,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [27,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [28,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [29,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [30,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [31,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [32,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [33,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [34,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [35,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [36,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [37,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [38,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [39,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [40,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>  [2,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#>  [3,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#>  [4,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#>  [5,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#>  [6,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#>  [7,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#>  [8,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [9,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [10,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [11,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [12,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [13,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [14,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [15,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [16,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [17,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [18,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [19,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [20,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [21,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [22,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [23,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [24,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [25,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [26,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [27,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [28,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [29,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [30,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [31,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [32,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [33,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [34,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [35,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [36,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [37,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [38,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [39,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [40,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>  [2,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#>  [3,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#>  [4,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#>  [5,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#>  [6,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#>  [7,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#>  [8,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [9,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [10,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [11,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [12,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [13,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [14,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [15,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [16,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [17,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [18,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [19,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [20,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [21,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [22,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [23,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [24,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [25,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [26,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [27,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [28,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [29,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [30,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [31,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [32,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [33,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [34,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [35,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [36,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [37,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [38,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [39,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [40,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#>  [2,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#>  [3,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#>  [4,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#>  [5,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#>  [6,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#>  [7,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#>  [8,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [9,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [10,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [11,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [12,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [13,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [14,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [15,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [16,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [17,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [18,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [19,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [20,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [21,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [22,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [23,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [24,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [25,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [26,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [27,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [28,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [29,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [30,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [31,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [32,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [33,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [34,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [35,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [36,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [37,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [38,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [39,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [40,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5038918 0.4420029 0.6919194 0.4772768 0.5619452 0.6448445 0.8444706
#>  [2,] 0.6973175 0.6377602 0.7846051 0.6436642 0.7628672 0.7730891 0.9155980
#>  [3,] 0.6020885 0.5101535 0.7593884 0.5707154 0.7263285 0.7140368 0.8902388
#>  [4,] 0.6544181 0.6132780 0.7915464 0.6481162 0.6979220 0.7687478 0.8888506
#>  [5,] 0.7118107 0.6652714 0.7860070 0.6311358 0.7976281 0.7663974 0.9298943
#>  [6,] 0.7495408 0.7093296 0.8195078 0.6843264 0.8279590 0.8092469 0.9456089
#>  [7,] 0.4440085 0.4077281 0.7115371 0.4778617 0.5665596 0.6052481 0.8423498
#>  [8,] 0.4690214 0.4346692 0.6684413 0.4825664 0.5842466 0.6201497 0.8394122
#>  [9,] 0.6456381 0.5703412 0.7480345 0.6055989 0.7308028 0.7428301 0.9029664
#> [10,] 0.4813780 0.4115986 0.7280618 0.5022523 0.5657445 0.6370533 0.8268201
#> [11,] 0.4842053 0.4083920 0.6994219 0.4894255 0.5445289 0.6038308 0.7951502
#> [12,] 0.4816811 0.4277116 0.7160946 0.5031542 0.5374437 0.6505650 0.8149640
#> [13,] 0.7095859 0.6479527 0.7963300 0.6802310 0.7798703 0.7859690 0.9165094
#> [14,] 0.6901129 0.6558295 0.8299873 0.6913345 0.7730399 0.8120368 0.9388002
#> [15,] 0.3986075 0.3152857 0.6852341 0.4133301 0.5158391 0.5591705 0.8047945
#> [16,] 0.7089866 0.6780304 0.8239129 0.7008789 0.7787079 0.8130034 0.9356701
#> [17,] 0.7164845 0.6599252 0.7971432 0.6478879 0.8038063 0.7818253 0.9344511
#> [18,] 0.5955659 0.5726847 0.7138164 0.5599557 0.6811828 0.6855821 0.8789782
#> [19,] 0.5959797 0.5614630 0.7007300 0.5408892 0.6967051 0.6640589 0.8897468
#> [20,] 0.5490074 0.5058192 0.7356079 0.5190277 0.6681029 0.7066890 0.9093138
#> [21,] 0.6273646 0.5820991 0.7305550 0.5561218 0.7008617 0.7105906 0.8847709
#> [22,] 0.6588610 0.5958621 0.7627169 0.5996530 0.6776445 0.7360263 0.8747375
#> [23,] 0.5866310 0.5547726 0.7108317 0.5484877 0.6940766 0.6688308 0.8896256
#> [24,] 0.6199918 0.5884718 0.7493570 0.5903115 0.6488385 0.7286073 0.8849275
#> [25,] 0.6155458 0.5390157 0.7400639 0.5377240 0.7502506 0.6946724 0.9111624
#> [26,] 0.4442383 0.4095259 0.6329730 0.4403323 0.5054101 0.5699407 0.7957706
#> [27,] 0.6715969 0.6420176 0.7708519 0.6232909 0.7435078 0.7557438 0.8990613
#> [28,] 0.7085745 0.6711156 0.8214888 0.6893875 0.7841468 0.8083291 0.9358794
#> [29,] 0.5601604 0.5379801 0.7314752 0.5434850 0.6519424 0.7061565 0.8658871
#> [30,] 0.6971905 0.6375231 0.7825341 0.6446136 0.7694775 0.7698381 0.9172001
#> [31,] 0.5649354 0.5176852 0.7308164 0.5155529 0.6931152 0.6776533 0.9067874
#> [32,] 0.5925682 0.5281280 0.7078563 0.5039482 0.7175849 0.6398565 0.8899122
#> [33,] 0.6615092 0.6080938 0.7526162 0.5963854 0.7685108 0.7246118 0.9189676
#> [34,] 0.6459304 0.5522173 0.7793293 0.5345169 0.7478269 0.7135863 0.9086996
#> [35,] 0.7204097 0.6676866 0.8168015 0.6448235 0.8201749 0.7873781 0.9449939
#> [36,] 0.6112225 0.5438126 0.7347289 0.5476484 0.7284544 0.7000878 0.9126295
#> [37,] 0.7019422 0.6524345 0.7993791 0.6499627 0.7933366 0.7794772 0.9335420
#> [38,] 0.6142107 0.5170180 0.7821130 0.5714181 0.7298127 0.7121726 0.9086946
#> [39,] 0.5038123 0.4050608 0.6569729 0.4054177 0.6403739 0.5881542 0.7949107
#> [40,] 0.6204576 0.5724230 0.7493374 0.5947542 0.6994359 0.7235168 0.9019229
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6699478 0.7337787 0.2841829
#>  [2,] 0.7393355 0.8661712 0.5275066
#>  [3,] 0.7595028 0.8282725 0.4228530
#>  [4,] 0.7179008 0.8162728 0.4296722
#>  [5,] 0.8053331 0.8784458 0.5810335
#>  [6,] 0.8206218 0.8996738 0.5886741
#>  [7,] 0.6439967 0.7456870 0.3000893
#>  [8,] 0.6494802 0.7389476 0.3154782
#>  [9,] 0.7550184 0.8250328 0.4952500
#> [10,] 0.6493954 0.7454547 0.2454509
#> [11,] 0.6015684 0.7161350 0.3062741
#> [12,] 0.6741954 0.6955143 0.2688945
#> [13,] 0.7504443 0.8629756 0.5371692
#> [14,] 0.7743654 0.8693113 0.4724726
#> [15,] 0.6571804 0.6859328 0.1971982
#> [16,] 0.7757655 0.8652974 0.5038494
#> [17,] 0.7948671 0.8798115 0.5885825
#> [18,] 0.7113649 0.7924320 0.4806210
#> [19,] 0.7272936 0.8124008 0.4163706
#> [20,] 0.7502989 0.8064919 0.3796427
#> [21,] 0.7491679 0.8035333 0.5027309
#> [22,] 0.7105144 0.8043681 0.4434306
#> [23,] 0.7306287 0.8107348 0.4140684
#> [24,] 0.7075911 0.7751044 0.4085014
#> [25,] 0.7676565 0.8426627 0.4601929
#> [26,] 0.6279329 0.6669166 0.2758007
#> [27,] 0.7612790 0.8435807 0.5194670
#> [28,] 0.7767970 0.8691215 0.5143076
#> [29,] 0.7176359 0.7785106 0.4007016
#> [30,] 0.7407341 0.8661897 0.5359853
#> [31,] 0.7458474 0.8068182 0.4243685
#> [32,] 0.7432598 0.8233123 0.4405259
#> [33,] 0.7643146 0.8603512 0.5047311
#> [34,] 0.7688188 0.8631009 0.4307865
#> [35,] 0.8203262 0.8999174 0.5704423
#> [36,] 0.7618940 0.8318441 0.4580731
#> [37,] 0.7697839 0.8828940 0.5217231
#> [38,] 0.7474262 0.8504917 0.4192987
#> [39,] 0.7139357 0.7544426 0.3423481
#> [40,] 0.7189863 0.8217195 0.4536144
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.7061630 0.2938370
#>  [2,] 0.7357219 0.2642781
#>  [3,] 0.7273190 0.2726810
#>  [4,] 0.7277083 0.2722917
#>  [5,] 0.7409641 0.2590359
#>  [6,] 0.7460787 0.2539213
#>  [7,] 0.7045483 0.2954517
#>  [8,] 0.7048602 0.2951398
#>  [9,] 0.7299954 0.2700046
#> [10,] 0.7051051 0.2948949
#> [11,] 0.7001347 0.2998653
#> [12,] 0.7031732 0.2968268
#> [13,] 0.7371453 0.2628547
#> [14,] 0.7385697 0.2614303
#> [15,] 0.6945645 0.3054355
#> [16,] 0.7392320 0.2607680
#> [17,] 0.7417585 0.2582415
#> [18,] 0.7216710 0.2783290
#> [19,] 0.7212892 0.2787108
#> [20,] 0.7227300 0.2772700
#> [21,] 0.7264962 0.2735038
#> [22,] 0.7248689 0.2751311
#> [23,] 0.7214818 0.2785182
#> [24,] 0.7208091 0.2791909
#> [25,] 0.7292956 0.2707044
#> [26,] 0.6933998 0.3066002
#> [27,] 0.7334355 0.2665645
#> [28,] 0.7396708 0.2603292
#> [29,] 0.7194323 0.2805677
#> [30,] 0.7359378 0.2640622
#> [31,] 0.7230880 0.2769120
#> [32,] 0.7229556 0.2770444
#> [33,] 0.7332602 0.2667398
#> [34,] 0.7317989 0.2682011
#> [35,] 0.7445784 0.2554216
#> [36,] 0.7279273 0.2720727
#> [37,] 0.7390915 0.2609085
#> [38,] 0.7289425 0.2710575
#> [39,] 0.7090253 0.2909747
#> [40,] 0.7259680 0.2740320
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]      [,3]         [,4]       [,5]      [,6]
#>  [1,]  0.01556769 -0.23303743 0.8091079 -0.090955533 0.24906024 0.5964535
#>  [2,]  0.83455630  0.56565590 1.2927074  0.591303632 1.16846356 1.2258370
#>  [3,]  0.41417475  0.04061966 1.1493292  0.284770697 0.97607375 0.9150713
#>  [4,]  0.63851659  0.46111228 1.3342721  0.610768962 0.83742205 1.2012542
#>  [5,]  0.90419461  0.68687497 1.3010222  0.537092637 1.37153540 1.1880792
#>  [6,]  1.09616481  0.89213056 1.5130165  0.773726153 1.57123097 1.4451240
#>  [7,] -0.22490925 -0.37336513 0.9028613 -0.088611216 0.26782817 0.4273811
#>  [8,] -0.12407337 -0.26282560 0.7011439 -0.069762556 0.34023089 0.4901838
#>  [9,]  0.59992031  0.28324343 1.0881567  0.428849296 0.99869971 1.0607303
#> [10,] -0.07452254 -0.35736089 0.9848109  0.009009176 0.26450956 0.5625974
#> [11,] -0.06319966 -0.37061663 0.8445464 -0.042304490 0.17858882 0.4214526
#> [12,] -0.07330839 -0.29119412 0.9251712  0.012617023 0.15005575 0.6215237
#> [13,]  0.89337351  0.61005200 1.3635125  0.754833367 1.26491090 1.3007966
#> [14,]  0.80064740  0.64476376 1.5855374  0.806365655 1.22555667 1.4632992
#> [15,] -0.41127055 -0.77552234 0.7779313 -0.350215600 0.06337762 0.2377961
#> [16,]  0.89046752  0.74473496 1.5430860  0.851486649 1.25815231 1.4696452
#> [17,]  0.92708971  0.66296104 1.3685339  0.609768052 1.41025586 1.2763353
#> [18,]  0.38702314  0.29281318 0.9139924  0.240982263 0.75921297 0.7795452
#> [19,]  0.38874153  0.24710176 0.8507763  0.163923025 0.83165678 0.6814351
#> [20,]  0.19666083  0.02327770 1.0232640  0.076147751 0.69961709 0.8793570
#> [21,]  0.52092720  0.33139645 0.9974401  0.225437107 0.85140470 0.8982539
#> [22,]  0.65822272  0.38825331 1.1676328  0.404019461 0.74296787 1.0254164
#> [23,]  0.35005532  0.21997308 0.8994268  0.194562112 0.81924764 0.7029016
#> [24,]  0.48951328  0.35765127 1.0951861  0.365253351 0.61393768 0.9875681
#> [25,]  0.47068457  0.15638062 1.0463007  0.151183434 1.09994918 0.8220552
#> [26,] -0.22397835 -0.36592564 0.5449924 -0.239813363 0.02164142 0.2816092
#> [27,]  0.71541627  0.58413195 1.2131280  0.503540248 1.06428087 1.1294848
#> [28,]  0.88847080  0.71323504 1.5264671  0.797257354 1.28999742 1.4391892
#> [29,]  0.24181320  0.15221341 1.0021199  0.174380413 0.62758810 0.8767897
#> [30,]  0.83395471  0.56462990 1.2804952  0.595445339 1.20536344 1.2073973
#> [31,]  0.26121700  0.07077032 0.9987688  0.062231510 0.81472365 0.7430079
#> [32,]  0.37459229  0.11263087 0.8849953  0.015793033 0.93251321 0.5747412
#> [33,]  0.67002700  0.43930654 1.1126145  0.390426715 1.19992127 0.9674541
#> [34,]  0.60119844  0.20963339 1.2617620  0.138287676 1.08705571 0.9128663
#> [35,]  0.94649460  0.69774023 1.4948260  0.596361658 1.51753295 1.3091932
#> [36,]  0.45245377  0.17570105 1.0187493  0.191173608 0.98679487 0.8477162
#> [37,]  0.85656359  0.62975771 1.3824181  0.618875099 1.34515638 1.2626222
#> [38,]  0.46504652  0.06809814 1.2780226  0.287639301 0.99367231 0.9059592
#> [39,]  0.01524961 -0.38442206 0.6498332 -0.382941129 0.57698748 0.3563400
#> [40,]  0.49149105  0.29174371 1.0950814  0.383654417 0.84461301 0.9619736
#>           [,7]      [,8]      [,9]       [,10]
#>  [1,] 1.691875 0.7079491 1.0138793 -0.92380666
#>  [2,] 2.383986 1.0425175 1.8675209  0.11013746
#>  [3,] 2.093183 1.1499553 1.5734334 -0.31107252
#>  [4,] 2.079053 0.9340729 1.4912968 -0.28318854
#>  [5,] 2.585067 1.4199660 1.9777939  0.32701759
#>  [6,] 2.855628 1.5205661 2.1936050  0.35848679
#>  [7,] 1.675817 0.5927534 1.0757400 -0.84687276
#>  [8,] 1.653861 0.6167551 1.0405060 -0.77463081
#>  [9,] 2.230628 1.1255589 1.5508248 -0.01900072
#> [10,] 1.563256 0.6163827 1.0745156 -1.12302330
#> [11,] 1.356254 0.4120045 0.9253699 -0.81759665
#> [12,] 1.482593 0.7272223 0.8260275 -1.00023872
#> [13,] 2.395838 1.1009835 1.8402275  0.14895150
#> [14,] 2.730459 1.2331270 1.8948830 -0.11022118
#> [15,] 1.416534 0.6507542 0.7811727 -1.40389844
#> [16,] 2.677239 1.2411578 1.8600039  0.01539779
#> [17,] 2.657162 1.3545168 1.9906466  0.35810874
#> [18,] 1.982790 0.9020224 1.3396479 -0.07755477
#> [19,] 2.088157 0.9809344 1.4656863 -0.33769055
#> [20,] 2.305285 1.1002070 1.4273745 -0.49106505
#> [21,] 2.038407 1.0941793 1.4085257  0.01092384
#> [22,] 1.943512 0.8978838 1.4138220 -0.22725038
#> [23,] 2.086922 0.9978145 1.4547917 -0.34717170
#> [24,] 2.039943 0.8837134 1.2373616 -0.37016417
#> [25,] 2.327912 1.1951252 1.6781746 -0.15956623
#> [26,] 1.360067 0.5233589 0.6942720 -0.96538809
#> [27,] 2.186837 1.1597045 1.6851150  0.07790747
#> [28,] 2.680721 1.2470972 1.8932134  0.05724602
#> [29,] 1.865072 0.9327651 1.2570076 -0.40254261
#> [30,] 2.404899 1.0497877 1.8676807  0.14419031
#> [31,] 2.275026 1.0765860 1.4294663 -0.30486550
#> [32,] 2.089845 1.0629809 1.5389519 -0.23902805
#> [33,] 2.428402 1.1764815 1.8182097  0.01892486
#> [34,] 2.297859 1.2016533 1.8412878 -0.27864282
#> [35,] 2.843734 1.5185589 2.1963069  0.28365587
#> [36,] 2.346172 1.1630912 1.5987536 -0.16810228
#> [37,] 2.642416 1.2070915 2.0201255  0.08694722
#> [38,] 2.297799 1.0849321 1.7384626 -0.32565286
#> [39,] 1.354784 0.9145762 1.1224487 -0.65284760
#> [40,] 2.218775 0.9394390 1.5280410 -0.18607766
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3016841 -0.5751366
#>  [2,] 0.3427657 -0.6810844
#>  [3,] 0.3311140 -0.6499484
#>  [4,] 0.3316241 -0.6514024
#>  [5,] 0.3517952 -0.6991901
#>  [6,] 0.3604068 -0.7173997
#>  [7,] 0.2947085 -0.5743429
#>  [8,] 0.2968772 -0.5736731
#>  [9,] 0.3348999 -0.6596993
#> [10,] 0.2985599 -0.5731681
#> [11,] 0.2878961 -0.5600433
#> [12,] 0.2960089 -0.5664455
#> [13,] 0.3446382 -0.6865452
#> [14,] 0.3486055 -0.6899425
#> [15,] 0.2827240 -0.5388225
#> [16,] 0.3493459 -0.6926347
#> [17,] 0.3518162 -0.7033121
#> [18,] 0.3206028 -0.6321627
#> [19,] 0.3233826 -0.6274830
#> [20,] 0.3254295 -0.6326146
#> [21,] 0.3289672 -0.6479505
#> [22,] 0.3277740 -0.6409689
#> [23,] 0.3232963 -0.6285276
#> [24,] 0.3216932 -0.6267853
#> [25,] 0.3335212 -0.6575306
#> [26,] 0.2813597 -0.5347027
#> [27,] 0.3401449 -0.6719786
#> [28,] 0.3495699 -0.6946881
#> [29,] 0.3187111 -0.6229362
#> [30,] 0.3426452 -0.6823158
#> [31,] 0.3235877 -0.6362432
#> [32,] 0.3242552 -0.6349147
#> [33,] 0.3399310 -0.6712961
#> [34,] 0.3385488 -0.6652197
#> [35,] 0.3574407 -0.7124621
#> [36,] 0.3321176 -0.6520144
#> [37,] 0.3484824 -0.6927698
#> [38,] 0.3331089 -0.6561549
#> [39,] 0.3041226 -0.5865324
#> [40,] 0.3284521 -0.6458088
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.7061630  0.7061630
#>  [2,]  0.2642781 -0.2642781
#>  [3,] -0.7273190  0.7273190
#>  [4,] -0.7277083  0.7277083
#>  [5,] -0.7409641  0.7409641
#>  [6,] -0.7460787  0.7460787
#>  [7,]  0.2954517 -0.2954517
#>  [8,]  0.2951398 -0.2951398
#>  [9,] -0.7299954  0.7299954
#> [10,]  0.2948949 -0.2948949
#> [11,]  0.2998653 -0.2998653
#> [12,] -0.7031732  0.7031732
#> [13,]  0.2628547 -0.2628547
#> [14,] -0.7385697  0.7385697
#> [15,] -0.6945645  0.6945645
#> [16,] -0.7392320  0.7392320
#> [17,] -0.7417585  0.7417585
#> [18,]  0.2783290 -0.2783290
#> [19,] -0.7212892  0.7212892
#> [20,] -0.7227300  0.7227300
#> [21,] -0.7264962  0.7264962
#> [22,] -0.7248689  0.7248689
#> [23,] -0.7214818  0.7214818
#> [24,] -0.7208091  0.7208091
#> [25,]  0.2707044 -0.2707044
#> [26,]  0.3066002 -0.3066002
#> [27,] -0.7334355  0.7334355
#> [28,] -0.7396708  0.7396708
#> [29,] -0.7194323  0.7194323
#> [30,]  0.2640622 -0.2640622
#> [31,]  0.2769120 -0.2769120
#> [32,]  0.2770444 -0.2770444
#> [33,]  0.2667398 -0.2667398
#> [34,] -0.7317989  0.7317989
#> [35,] -0.7445784  0.7445784
#> [36,]  0.2720727 -0.2720727
#> [37,]  0.2609085 -0.2609085
#> [38,]  0.2710575 -0.2710575
#> [39,] -0.7090253  0.7090253
#> [40,]  0.2740320 -0.2740320
#> 
#> $L
#> [1] 0.6928638 0.7238526 0.7146484 0.7327053 0.7315227 0.8849826
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4637681