Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.2461609 -0.18891105 -0.06318084 -0.18058001 -0.015600178 -0.01703421
#>  [2,] -0.2344067 -0.09605864 -0.11060486 -0.17905234 -0.088738158 -0.02136480
#>  [3,] -0.3909513 -0.10883982 -0.01780914  0.03858534 -0.035085607 -0.16064780
#>  [4,] -0.2694715 -0.07407061 -0.16311731 -0.03561935 -0.081296135 -0.01900093
#>  [5,] -0.3539972 -0.12942179 -0.02890051 -0.05827014 -0.064143107 -0.16440910
#>  [6,] -0.3652134 -0.10156371 -0.13624066 -0.03472253 -0.046274599 -0.01316466
#>  [7,] -0.3012344 -0.04761343 -0.10135445 -0.05563813 -0.089358015  0.06596235
#>  [8,] -0.2704846 -0.14492572 -0.10393220 -0.10336963 -0.005819102 -0.13445087
#>  [9,] -0.2531701 -0.10975951 -0.01378223 -0.13781644 -0.078268966 -0.08010329
#> [10,] -0.3486907 -0.07555401 -0.12945568 -0.20319508 -0.153239550 -0.11295357
#>               V15         V16          V17         V18         V19         V2
#>  [1,] -0.01874104 -0.04047442 -0.037891944 -0.01962069 -0.07361335 -0.2447492
#>  [2,] -0.03703662  0.03349769  0.053663852 -0.01028388  0.02329444 -0.3035731
#>  [3,] -0.07244487 -0.07610276 -0.062725541 -0.01887450 -0.03978117 -0.4117875
#>  [4,] -0.14626046 -0.19576351 -0.023489857 -0.05613780  0.04987164 -0.1919753
#>  [5,] -0.03690191  0.03128050 -0.113658208  0.04292401  0.04614982 -0.2773747
#>  [6,] -0.11890318 -0.07400625  0.063636375  0.06113814 -0.04085751 -0.1684810
#>  [7,]  0.09535594  0.03103617  0.087774275  0.10792779  0.19095064 -0.3882601
#>  [8,] -0.05130014  0.01385043  0.006761489 -0.02911843 -0.03154213 -0.3066855
#>  [9,]  0.02781702  0.11272880  0.055177360  0.10777395  0.16777598 -0.3127092
#> [10,] -0.13109073 -0.11925962 -0.194868563 -0.09763062 -0.10938149 -0.2573336
#>               V20          V21         V22         V23        V24          V25
#>  [1,] -0.01277779  0.035363107 -0.04094615  0.12840187 0.02257484  0.069886673
#>  [2,]  0.09706256  0.008462132  0.16469616  0.13675159 0.05758870  0.149988883
#>  [3,]  0.15539169  0.221176741  0.16809022  0.21574721 0.16075869  0.087130041
#>  [4,] -0.06899665 -0.022101980  0.06012681  0.05814383 0.03291309  0.012054408
#>  [5,] -0.11583043 -0.084032762  0.03273342 -0.03898518 0.10753753  0.189615915
#>  [6,]  0.05326827  0.083484598  0.14759570  0.14494112 0.14508241  0.227684834
#>  [7,]  0.16572181  0.234207247  0.13757645  0.16633874 0.32119374  0.261956943
#>  [8,] -0.03917098  0.002138427  0.19823128  0.07499105 0.18913535  0.184835884
#>  [9,]  0.16329356  0.157729521  0.07839554  0.21929709 0.16218883  0.162350976
#> [10,] -0.12304122 -0.075285519 -0.01191354  0.09190044 0.05749366 -0.007669456
#>              V26        V27         V28         V29         V3         V30
#>  [1,] 0.11211487 0.18958978 0.221933464  0.14257349 -0.2186219  0.18683621
#>  [2,] 0.14540339 0.16679015 0.005107946 -0.05302857 -0.3643409  0.06483264
#>  [3,] 0.12402962 0.12339079 0.038835392  0.13643634 -0.2387703  0.13490140
#>  [4,] 0.09291944 0.14118246 0.007452144  0.11690911 -0.2280019  0.09089613
#>  [5,] 0.11424034 0.01167992 0.046526171  0.07724731 -0.3659326  0.03566316
#>  [6,] 0.22013957 0.20880056 0.229102333  0.13206573 -0.3524146  0.07255438
#>  [7,] 0.22192075 0.32982901 0.239460443  0.24707568 -0.3523495  0.21843860
#>  [8,] 0.27376622 0.24367596 0.139562070  0.13069458 -0.3146341  0.08685968
#>  [9,] 0.03607281 0.01378076 0.178136060 -0.03177095 -0.3169350 -0.02358786
#> [10,] 0.02707609 0.14238481 0.035663939  0.08846160 -0.2206883  0.01974297
#>               V31         V32          V33         V34           V35
#>  [1,]  0.06760598  0.01228320 -0.004403426  0.03423351 -0.0179880354
#>  [2,] -0.01709826 -0.21245710 -0.158995721 -0.02482929 -0.0188388329
#>  [3,]  0.11129390  0.12759664  0.089401016  0.12723711  0.0655412651
#>  [4,]  0.12250257  0.08340879 -0.005963142  0.02052204 -0.1543109249
#>  [5,]  0.14517772  0.04896576  0.012990406  0.03810215 -0.0002148005
#>  [6,]  0.14281701 -0.01128279 -0.022017323 -0.01286309  0.0121858575
#>  [7,]  0.13458069  0.18700851  0.222962965  0.04654739  0.1615409383
#>  [8,]  0.12210816 -0.00980747 -0.057958238  0.09411646  0.1286420521
#>  [9,] -0.08223892  0.02545074  0.111193214 -0.02765443 -0.0617449804
#> [10,] -0.04994665  0.07216509  0.065970210 -0.05591977 -0.1486176637
#>               V36          V37          V38         V39         V4         V40
#>  [1,] -0.09076465 -0.109109457 -0.183790332 -0.04144992 -0.2539085 -0.03127078
#>  [2,] -0.01198763 -0.135093215 -0.072683554 -0.05025439 -0.2239147 -0.06341459
#>  [3,] -0.02441193  0.067379055 -0.035858372  0.02293310 -0.2354627 -0.01799510
#>  [4,] -0.17200352 -0.111758984 -0.018919912 -0.14276489 -0.2192666 -0.11464374
#>  [5,]  0.01547913 -0.050852677  0.006425548 -0.09163400 -0.3363285 -0.02744523
#>  [6,]  0.00204029  0.002415396 -0.096985198 -0.10968568 -0.2639429 -0.09760485
#>  [7,]  0.06869426  0.066578859 -0.033725375 -0.04343509 -0.2178593  0.03778594
#>  [8,]  0.15314659  0.001403546  0.078127715 -0.11307741 -0.3054783 -0.04794365
#>  [9,]  0.11269099  0.068351480 -0.052960303  0.01576056 -0.2796214 -0.09148051
#> [10,] -0.10500681  0.066287120 -0.106623235  0.02121440 -0.3591658 -0.03933709
#>                V41          V42         V43          V44         V45
#>  [1,] -0.137596412 -0.070362489 -0.18943020 -0.239238783 -0.16390739
#>  [2,] -0.191771882 -0.121977505 -0.17741181 -0.217064521 -0.13389808
#>  [3,] -0.074377616  0.027332614 -0.03825322 -0.104393937 -0.13287842
#>  [4,] -0.147501086 -0.156623728 -0.01824937 -0.091149578 -0.16647921
#>  [5,] -0.010956027 -0.008107096 -0.13570691 -0.124857713 -0.13420555
#>  [6,]  0.000356308 -0.037836098 -0.04964818 -0.197325227 -0.05408824
#>  [7,]  0.055831723 -0.044738858 -0.10390923 -0.006719927 -0.09986364
#>  [8,]  0.042524653 -0.051236865 -0.06465449 -0.075845394 -0.04112395
#>  [9,] -0.041787133 -0.122884358 -0.03267910 -0.015031723 -0.14019936
#> [10,] -0.159061369 -0.075835293 -0.03281166 -0.021137606 -0.16510418
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.12040535 -0.2573465 -0.2167174 -0.3579475 -0.3422679 -0.3900869
#>  [2,] -0.29870750 -0.2803143 -0.1876063 -0.2500813 -0.1502453 -0.3032675
#>  [3,] -0.24533730 -0.1299818 -0.1920640 -0.2296485 -0.2447606 -0.4367056
#>  [4,] -0.23683760 -0.2760699 -0.2244909 -0.1695315 -0.2020776 -0.2420821
#>  [5,] -0.04958241 -0.1505815 -0.1056379 -0.3197697 -0.1835631 -0.2942506
#>  [6,] -0.08907853 -0.1046406 -0.1630421 -0.1835907 -0.2446662 -0.2624063
#>  [7,] -0.06656276 -0.1390971 -0.2766195 -0.3524276 -0.2818455 -0.3305432
#>  [8,] -0.17988316 -0.1015735 -0.2925311 -0.2238139 -0.2745922 -0.3072301
#>  [9,] -0.19738828 -0.2543376 -0.2212657 -0.3298022 -0.2265404 -0.3940665
#> [10,] -0.17843562 -0.1091652 -0.1097283 -0.2288853 -0.2734139 -0.3841355
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3269517 -0.2783542 -0.2505506 -0.2640697 -0.2866925 -0.3635580
#>  [2,] -0.3963518 -0.3622419 -0.3155082 -0.2397157 -0.4178242 -0.2746376
#>  [3,] -0.4128505 -0.3112753 -0.4704080 -0.4544842 -0.4238961 -0.3277776
#>  [4,] -0.2140224 -0.3912329 -0.2559967 -0.2908395 -0.3343483 -0.3487139
#>  [5,] -0.2993099 -0.3020876 -0.3554447 -0.4138899 -0.3095472 -0.3793910
#>  [6,] -0.3082486 -0.3575266 -0.3115059 -0.3612273 -0.2346547 -0.3316274
#>  [7,] -0.4176054 -0.3148552 -0.3746309 -0.3859274 -0.3608522 -0.3043460
#>  [8,] -0.3641878 -0.3403166 -0.2553207 -0.2953884 -0.3286148 -0.3301627
#>  [9,] -0.2489549 -0.3844874 -0.3963910 -0.2581915 -0.3163032 -0.4263802
#> [10,] -0.2600346 -0.2864901 -0.2943573 -0.3660026 -0.3653776 -0.2488475
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.2899188 -0.2634347 -0.3918990 -0.15444224 -0.3879154 -0.17854017
#>  [2,] -0.2781210 -0.3697885 -0.3424423 -0.19230264 -0.3702934 -0.10982304
#>  [3,] -0.3200118 -0.4005228 -0.4170527 -0.23706443 -0.2933596 -0.26752567
#>  [4,] -0.3763250 -0.3488197 -0.2217460 -0.18454482 -0.3123796 -0.10677509
#>  [5,] -0.2906652 -0.4169545 -0.2815853 -0.28004209 -0.4167150 -0.21757221
#>  [6,] -0.3226636 -0.3703480 -0.2033358 -0.09984775 -0.2395331 -0.18479766
#>  [7,] -0.4198387 -0.4570240 -0.3959535 -0.18088250 -0.4470814 -0.17195834
#>  [8,] -0.3744715 -0.2455253 -0.2663858 -0.24955845 -0.3931180 -0.20329232
#>  [9,] -0.3719321 -0.2686822 -0.4436723 -0.23309593 -0.2885260 -0.09500939
#> [10,] -0.3871627 -0.3822761 -0.3288611 -0.15258836 -0.2794191 -0.29273331
#>               V8           V9
#>  [1,] -0.1359633 -0.223715075
#>  [2,] -0.1243334 -0.115108066
#>  [3,] -0.2655573 -0.218244077
#>  [4,] -0.1735635 -0.155257898
#>  [5,] -0.1949955 -0.171763788
#>  [6,] -0.1241326 -0.217792641
#>  [7,] -0.2135989  0.003838332
#>  [8,] -0.1227390 -0.133299529
#>  [9,] -0.1596937 -0.027169300
#> [10,] -0.1033684 -0.193545668
#> 
#> $W[[2]]
#>              [,1]        [,2]        [,3]         [,4]        [,5]        [,6]
#> [1,] -0.105347049 -0.08514136  0.04694715  0.074450958  0.03061422  0.08840158
#> [2,]  0.003931523  0.09514317 -0.04124134 -0.001379022 -0.03709994 -0.03546225
#>             [,7]       [,8]         [,9]        [,10]
#> [1,]  0.13439635 0.11529035  0.007519917 -0.066098682
#> [2,] -0.09167953 0.05014556 -0.095787013 -0.008746565
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.694549e-04  1.467794e-03  2.130782e-03  0.0024381161  2.230604e-03
#>  [2,]  2.598652e-04  2.225650e-03  3.241067e-03  0.0037176337  3.397715e-03
#>  [3,]  7.427883e-06  5.638807e-05  8.946847e-05  0.0001084029  1.013715e-04
#>  [4,] -4.417894e-05 -4.108599e-04 -5.470861e-04 -0.0005777548 -5.229726e-04
#>  [5,] -3.933472e-05 -3.399314e-04 -4.263526e-04 -0.0004250353 -3.934057e-04
#>  [6,] -4.581893e-05 -3.682067e-04 -5.015976e-04 -0.0005353123 -5.009135e-04
#>  [7,] -2.911059e-05 -2.076633e-04 -2.991022e-04 -0.0003311689 -3.168033e-04
#>  [8,]  5.484824e-06  4.348957e-05  8.997392e-05  0.0001259412  1.178700e-04
#>  [9,] -1.165061e-05 -9.454003e-05 -1.108439e-04 -0.0001050243 -9.683734e-05
#> [10,]  7.486515e-05  7.192232e-04  1.039150e-03  0.0011946364  1.065592e-03
#>                 V14           V15           V16           V17           V18
#>  [1,]  2.003668e-03  1.487923e-03  1.288536e-03  1.359289e-03  1.937368e-03
#>  [2,]  3.001591e-03  2.218284e-03  1.893414e-03  1.968964e-03  2.832463e-03
#>  [3,]  9.982899e-05  8.556799e-05  7.942158e-05  8.592088e-05  1.131082e-04
#>  [4,] -4.233706e-04 -2.543572e-04 -1.536295e-04 -1.210169e-04 -2.173349e-04
#>  [5,] -3.176256e-04 -1.847732e-04 -9.343005e-05 -4.391532e-05 -9.121231e-05
#>  [6,] -4.575243e-04 -3.388440e-04 -2.676506e-04 -2.467121e-04 -3.329181e-04
#>  [7,] -3.302608e-04 -3.012653e-04 -2.831039e-04 -2.892812e-04 -3.327395e-04
#>  [8,]  1.264029e-04  1.206316e-04  1.290976e-04  1.562204e-04  2.007413e-04
#>  [9,] -7.887905e-05 -4.883983e-05 -2.823611e-05 -1.232843e-05 -1.640875e-05
#> [10,]  9.389199e-04  6.825019e-04  6.037916e-04  6.425970e-04  9.386416e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  3.085939e-03  1.685846e-04  0.0035084936  0.0031295054  2.874063e-03
#>  [2,]  4.620421e-03  2.610224e-04  0.0052419402  0.0046796321  4.216358e-03
#>  [3,]  1.576423e-04  7.090195e-06  0.0001722101  0.0001491298  1.394764e-04
#>  [4,] -5.359446e-04 -4.680551e-05 -0.0006753049 -0.0006220773 -5.420220e-04
#>  [5,] -3.426319e-04 -4.256494e-05 -0.0004909542 -0.0004708987 -3.963613e-04
#>  [6,] -6.003229e-04 -4.696574e-05 -0.0007444534 -0.0006752938 -5.947438e-04
#>  [7,] -4.452247e-04 -2.985511e-05 -0.0005089491 -0.0004475734 -3.909099e-04
#>  [8,]  2.359076e-04  4.687461e-06  0.0002306515  0.0001944001  1.942532e-04
#>  [9,] -7.724313e-05 -1.261640e-05 -0.0001199022 -0.0001177081 -9.605658e-05
#> [10,]  1.481942e-03  7.401126e-05  0.0016542114  0.0014543582  1.332602e-03
#>                 V23           V24           V25           V26           V27
#>  [1,]  2.777074e-03  0.0024227158  2.263894e-03  2.373881e-03  0.0026160074
#>  [2,]  3.977741e-03  0.0034119170  3.137135e-03  3.317294e-03  0.0037068010
#>  [3,]  1.393118e-04  0.0001219305  1.186758e-04  1.187579e-04  0.0001282727
#>  [4,] -4.696618e-04 -0.0003682391 -3.215782e-04 -4.106289e-04 -0.0004973738
#>  [5,] -3.231607e-04 -0.0002456487 -2.165701e-04 -3.221483e-04 -0.0003867403
#>  [6,] -5.357872e-04 -0.0004388850 -4.046170e-04 -4.792776e-04 -0.0005349470
#>  [7,] -3.570351e-04 -0.0002981442 -2.776689e-04 -3.044065e-04 -0.0003378256
#>  [8,]  2.083662e-04  0.0001930240  1.879957e-04  1.630785e-04  0.0001675566
#>  [9,] -7.356999e-05 -0.0000548384 -4.845737e-05 -8.146684e-05 -0.0001010588
#> [10,]  1.309279e-03  0.0011541210  1.075745e-03  1.105583e-03  0.0012435739
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0024268080  1.665002e-03  2.683397e-04  8.167170e-04 -2.148629e-04
#>  [2,]  0.0034256358  2.274209e-03  4.096509e-04  1.002741e-03 -5.389071e-04
#>  [3,]  0.0001197627  8.874270e-05  1.171781e-05  4.706633e-05  4.682010e-06
#>  [4,] -0.0004634521 -2.713101e-04 -6.671598e-05 -1.099605e-04  1.362803e-04
#>  [5,] -0.0003683466 -2.256915e-04 -5.386948e-05 -1.318885e-04  6.088748e-05
#>  [6,] -0.0004934237 -3.207173e-04 -6.566127e-05 -1.766174e-04  5.171862e-05
#>  [7,] -0.0003087267 -2.116186e-04 -4.273381e-05 -1.268471e-04 -1.114054e-05
#>  [8,]  0.0001520297  1.178542e-04  1.127945e-05  5.388006e-05  1.004133e-05
#>  [9,] -0.0000989312 -6.081923e-05 -1.472588e-05 -4.508935e-05 -3.321550e-06
#> [10,]  0.0011681949  7.820326e-04  1.248721e-04  3.373076e-04 -1.782941e-04
#>                 V32           V33           V34           V35           V36
#>  [1,]  2.794730e-04  3.251763e-04 -2.917250e-04 -1.053331e-03 -1.354908e-03
#>  [2,]  2.166525e-04  2.724020e-04 -6.770375e-04 -1.845392e-03 -2.265441e-03
#>  [3,]  2.654670e-05  3.105626e-05  1.028675e-05 -1.899835e-05 -4.048465e-05
#>  [4,] -1.287004e-05  2.048761e-05  2.386481e-04  4.388441e-04  4.714314e-04
#>  [5,] -5.636090e-05 -3.193886e-05  1.465949e-04  3.000646e-04  3.232541e-04
#>  [6,] -8.368445e-05 -8.893142e-05  8.281434e-05  2.648738e-04  3.408447e-04
#>  [7,] -8.766638e-05 -9.702518e-05 -1.646043e-05  7.563276e-05  1.515497e-04
#>  [8,]  2.659336e-05  3.817891e-05  3.201070e-05  8.219318e-06 -2.271828e-05
#>  [9,] -3.234752e-05 -1.864578e-05  3.207868e-05  7.344082e-05  8.139171e-05
#> [10,]  5.110042e-05  5.876082e-05 -2.434201e-04 -6.302554e-04 -7.633973e-04
#>                 V37           V38           V39            V4           V40
#>  [1,] -7.128391e-04  1.005878e-04  5.966927e-04  4.730513e-04  5.571086e-04
#>  [2,] -1.189487e-03  1.278285e-04  9.214821e-04  7.099832e-04  8.071792e-04
#>  [3,] -1.637047e-05  5.655430e-06  1.798718e-05  2.180880e-05  2.329758e-05
#>  [4,]  2.495921e-04 -4.893238e-05 -2.070245e-04 -1.108020e-04 -1.060944e-04
#>  [5,]  1.506114e-04 -1.023783e-04 -2.152654e-04 -8.638640e-05 -9.069248e-05
#>  [6,]  1.659986e-04 -6.467721e-05 -1.702338e-04 -1.119887e-04 -1.017579e-04
#>  [7,]  5.412969e-05 -3.717150e-05 -6.661236e-05 -7.294517e-05 -4.482946e-05
#>  [8,] -1.219951e-05 -1.633014e-05 -6.936629e-06  2.291599e-05  3.013121e-05
#>  [9,]  3.395296e-05 -3.291786e-05 -6.083712e-05 -2.279675e-05 -2.388475e-05
#> [10,] -4.425883e-04 -4.993654e-05  2.136413e-04  2.229069e-04  2.383081e-04
#>                 V41           V42           V43           V44           V45
#>  [1,]  8.393054e-04  1.461628e-03  1.343274e-03  1.295611e-03  1.483972e-03
#>  [2,]  1.224508e-03  2.172180e-03  1.999465e-03  1.932695e-03  2.224222e-03
#>  [3,]  3.643578e-05  6.224666e-05  5.000243e-05  4.506568e-05  5.316865e-05
#>  [4,] -1.910412e-04 -3.381117e-04 -3.573686e-04 -3.869555e-04 -4.744196e-04
#>  [5,] -1.571175e-04 -2.697305e-04 -3.011713e-04 -3.292402e-04 -4.154635e-04
#>  [6,] -1.824202e-04 -3.289269e-04 -3.210953e-04 -3.206034e-04 -4.075022e-04
#>  [7,] -8.532029e-05 -1.768363e-04 -1.579159e-04 -1.378168e-04 -1.914401e-04
#>  [8,]  4.012107e-05  6.916318e-05  4.027934e-05  2.190768e-05  1.877669e-05
#>  [9,] -4.235787e-05 -6.954597e-05 -7.963642e-05 -9.054424e-05 -1.183159e-04
#> [10,]  3.665363e-04  6.693072e-04  6.141858e-04  6.061944e-04  6.836818e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.259782e-03  8.554209e-04  5.740466e-04  3.460593e-04  5.521337e-04
#>  [2,]  1.877198e-03  1.289300e-03  8.761420e-04  5.235710e-04  8.336705e-04
#>  [3,]  4.762886e-05  3.359865e-05  2.115269e-05  1.314642e-05  2.472663e-05
#>  [4,] -3.966712e-04 -2.682672e-04 -1.904601e-04 -1.086640e-04 -1.292853e-04
#>  [5,] -3.427370e-04 -2.237465e-04 -1.637186e-04 -9.126165e-05 -9.897482e-05
#>  [6,] -3.460345e-04 -2.248530e-04 -1.595838e-04 -9.223578e-05 -1.222249e-04
#>  [7,] -1.702805e-04 -1.092186e-04 -7.776718e-05 -4.560533e-05 -7.420510e-05
#>  [8,]  2.146489e-05  1.869814e-05  7.518129e-06  7.276745e-06  2.705811e-05
#>  [9,] -9.915768e-05 -6.440608e-05 -4.779817e-05 -2.610147e-05 -2.447446e-05
#> [10,]  5.862667e-04  4.084679e-04  2.683704e-04  1.615917e-04  2.622946e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.103074e-04  1.119250e-04  1.250389e-04  6.820932e-05  4.966304e-05
#>  [2,]  1.619394e-04  1.699559e-04  1.874439e-04  1.036628e-04  7.304553e-05
#>  [3,]  5.057672e-06  4.532217e-06  5.722002e-06  3.488582e-06  2.310171e-06
#>  [4,] -3.027309e-05 -3.488401e-05 -3.243539e-05 -1.405436e-05 -1.255484e-05
#>  [5,] -2.588323e-05 -3.084493e-05 -2.643248e-05 -1.092763e-05 -1.163177e-05
#>  [6,] -2.958006e-05 -3.176412e-05 -3.146599e-05 -1.504125e-05 -1.390820e-05
#>  [7,] -1.816356e-05 -1.784686e-05 -1.969601e-05 -1.059285e-05 -9.171013e-06
#>  [8,]  4.130870e-06  2.071747e-06  5.332252e-06  4.102859e-06  2.001375e-06
#>  [9,] -7.608600e-06 -8.972535e-06 -7.253801e-06 -2.685668e-06 -3.338992e-06
#> [10,]  4.950583e-05  4.999761e-05  5.832040e-05  3.088910e-05  2.047421e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  3.612995e-05  4.906090e-05  3.831220e-05  6.590436e-05  4.105002e-05
#>  [2,]  5.402625e-05  7.342324e-05  5.332619e-05  9.995787e-05  6.294984e-05
#>  [3,]  2.004527e-06  2.412930e-06  2.084481e-06  3.274420e-06  1.861383e-06
#>  [4,] -1.018123e-05 -1.106424e-05 -6.241356e-06 -1.509560e-05 -1.104124e-05
#>  [5,] -1.021551e-05 -9.168635e-06 -4.745761e-06 -1.213467e-05 -9.617135e-06
#>  [6,] -1.197326e-05 -1.203482e-05 -8.275719e-06 -1.598605e-05 -1.084539e-05
#>  [7,] -8.633082e-06 -7.877675e-06 -6.591965e-06 -1.091371e-05 -6.872141e-06
#>  [8,]  1.453767e-06  2.511862e-06  2.787366e-06  3.547004e-06  1.579660e-06
#>  [9,] -2.952853e-06 -2.386485e-06 -1.280669e-06 -3.050759e-06 -2.572464e-06
#> [10,]  1.315138e-05  2.133124e-05  1.833634e-05  2.900081e-05  1.737528e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  4.391787e-04  2.065389e-05  5.020097e-04  6.838643e-04  1.128028e-03
#>  [2,]  6.497875e-04  3.092301e-05  7.314535e-04  1.010991e-03  1.693100e-03
#>  [3,]  2.263023e-05  1.010213e-06  2.669270e-05  2.988458e-05  4.209282e-05
#>  [4,] -9.128893e-05 -4.642845e-06 -9.427433e-05 -1.656833e-04 -3.188726e-04
#>  [5,] -6.801135e-05 -4.324656e-06 -6.913185e-05 -1.372619e-04 -2.702311e-04
#>  [6,] -8.988230e-05 -5.485174e-06 -1.010622e-04 -1.623547e-04 -2.865016e-04
#>  [7,] -6.804249e-05 -4.206857e-06 -7.693410e-05 -1.049667e-04 -1.566100e-04
#>  [8,]  2.763013e-05  9.836793e-07  3.382198e-05  2.737111e-05  2.770283e-05
#>  [9,] -1.704094e-05 -1.307123e-06 -1.803762e-05 -3.849477e-05 -7.637266e-05
#> [10,]  2.109835e-04  8.637806e-06  2.471798e-04  3.376585e-04  5.530081e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.02761301 -0.03107697 -0.05010751 -0.02498881 -0.02754811 -0.04468501
#> [2,]  0.02761301  0.03107697  0.05010751  0.02498881  0.02754811  0.04468501
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.06579205 -0.04132874 -0.04996663 -0.01621142
#> [2,]  0.06579205  0.04132874  0.04996663  0.01621142
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.2412935 0.2403942 0.7637208 0.3205957 0.5323057 0.2919882 1.0365798
#>  [8] 0.4655308 0.6542714 0.3736699
#> 
#> $B[[2]]
#> [1]  0.04107497 -0.19054764
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  3.056820e-03  4.345764e-03  1.683835e-04 -4.569429e-04 -3.540331e-04
#>  [6] -6.164795e-04 -4.759669e-04  2.434728e-04 -9.187552e-05  1.377799e-03
#> 
#> $vB[[2]]
#> [1] -0.07167415  0.07167415
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>  [2,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#>  [3,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#>  [4,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [5,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#>  [6,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#>  [7,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#>  [8,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#>  [9,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [10,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [11,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [12,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [13,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [14,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [15,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [16,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [17,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [18,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [19,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [20,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [21,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [22,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [23,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [24,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [25,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [26,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [27,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [28,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [29,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [30,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [31,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [32,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [33,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [34,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [35,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [36,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [37,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [38,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [39,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [40,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>  [2,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#>  [3,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#>  [4,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [5,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#>  [6,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#>  [7,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#>  [8,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#>  [9,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [10,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [11,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [12,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [13,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [14,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [15,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [16,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [17,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [18,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [19,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [20,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [21,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [22,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [23,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [24,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [25,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [26,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [27,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [28,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [29,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [30,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [31,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [32,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [33,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [34,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [35,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [36,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [37,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [38,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [39,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [40,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>  [2,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#>  [3,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#>  [4,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [5,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#>  [6,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#>  [7,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#>  [8,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#>  [9,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [10,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [11,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [12,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [13,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [14,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [15,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [16,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [17,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [18,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [19,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [20,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [21,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [22,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [23,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [24,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [25,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [26,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [27,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [28,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [29,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [30,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [31,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [32,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [33,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [34,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [35,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [36,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [37,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [38,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [39,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [40,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>  [2,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#>  [3,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#>  [4,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [5,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#>  [6,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#>  [7,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#>  [8,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#>  [9,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [10,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [11,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [12,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [13,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [14,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [15,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [16,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [17,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [18,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [19,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [20,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [21,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [22,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [23,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [24,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [25,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [26,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [27,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [28,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [29,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [30,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [31,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [32,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [33,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [34,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [35,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [36,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [37,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [38,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [39,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [40,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>  [2,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#>  [3,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#>  [4,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [5,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#>  [6,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#>  [7,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#>  [8,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#>  [9,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [10,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [11,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [12,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [13,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [14,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [15,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [16,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [17,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [18,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [19,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [20,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [21,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [22,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [23,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [24,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [25,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [26,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [27,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [28,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [29,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [30,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [31,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [32,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [33,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [34,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [35,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [36,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [37,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [38,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [39,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [40,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#>  [2,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#>  [3,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#>  [4,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [5,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#>  [6,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#>  [7,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#>  [8,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#>  [9,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [10,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [11,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [12,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [13,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [14,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [15,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [16,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [17,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [18,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [19,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [20,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [21,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [22,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [23,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [24,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [25,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [26,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [27,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [28,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [29,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [30,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [31,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [32,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [33,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [34,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [35,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [36,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [37,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [38,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [39,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [40,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5063976 0.4381682 0.7019098 0.4371681 0.5634218 0.6075337 0.8967461
#>  [2,] 0.5968266 0.5960156 0.8132044 0.4987436 0.5773667 0.7516793 0.9611314
#>  [3,] 0.5404290 0.3796306 0.7922587 0.4626830 0.6203311 0.7165857 0.9536941
#>  [4,] 0.3922924 0.4575686 0.7426890 0.3602567 0.5096705 0.5793963 0.9150592
#>  [5,] 0.4716166 0.5058575 0.7522089 0.3754156 0.5534187 0.6803995 0.9567117
#>  [6,] 0.6822299 0.6282937 0.8863923 0.5942017 0.6921507 0.8253593 0.9745755
#>  [7,] 0.4795954 0.3906957 0.8373973 0.4105772 0.5796275 0.6837765 0.9592547
#>  [8,] 0.6690603 0.5901538 0.8465586 0.5982244 0.6607687 0.7963745 0.9609057
#>  [9,] 0.6457308 0.6665785 0.8622339 0.5725544 0.6514766 0.7920888 0.9661786
#> [10,] 0.6275400 0.5212607 0.8056839 0.5419062 0.6298632 0.7527519 0.9494127
#> [11,] 0.4389800 0.5311056 0.7770204 0.3763348 0.5035566 0.6827336 0.9394004
#> [12,] 0.4773109 0.5128074 0.8110940 0.4087771 0.5707989 0.6950795 0.9535790
#> [13,] 0.6312298 0.5483856 0.8652271 0.5448135 0.6758197 0.7815351 0.9669815
#> [14,] 0.4919544 0.3324288 0.7749201 0.4227647 0.5937027 0.6696028 0.9484959
#> [15,] 0.6318480 0.5999102 0.7888135 0.5608808 0.6705505 0.7515800 0.9281887
#> [16,] 0.3634115 0.3038246 0.7312393 0.3232417 0.5315234 0.5810206 0.9173532
#> [17,] 0.5757694 0.5427911 0.8511466 0.4768463 0.6269590 0.7749008 0.9664434
#> [18,] 0.5296941 0.3645037 0.7922653 0.4311691 0.6132327 0.6403663 0.9279401
#> [19,] 0.4455017 0.4553189 0.6672839 0.3910691 0.4858767 0.5654062 0.8996596
#> [20,] 0.6155089 0.4744960 0.8321602 0.5423705 0.6513589 0.7414487 0.9576762
#> [21,] 0.5692245 0.5808447 0.7998659 0.4848878 0.5778468 0.7458572 0.9578206
#> [22,] 0.5123872 0.5309065 0.8132649 0.4712205 0.5622936 0.7415455 0.9560704
#> [23,] 0.4636362 0.5421371 0.8058141 0.4032537 0.5358931 0.6801292 0.9475947
#> [24,] 0.5714146 0.5564641 0.8062664 0.4744831 0.5902626 0.7464180 0.9639372
#> [25,] 0.6185189 0.5949444 0.8375122 0.5381928 0.6210217 0.7924852 0.9637785
#> [26,] 0.5809348 0.4754057 0.8496586 0.4759977 0.6599957 0.7686971 0.9703188
#> [27,] 0.5551746 0.5652760 0.7735273 0.4895437 0.5879990 0.7122010 0.9387524
#> [28,] 0.6377368 0.5151347 0.8124355 0.5423533 0.6451021 0.7543076 0.9495076
#> [29,] 0.5672275 0.5191392 0.8248445 0.5044497 0.5856708 0.7128339 0.9553629
#> [30,] 0.5944458 0.6031460 0.8233944 0.5076134 0.6093958 0.7817807 0.9658668
#> [31,] 0.3830312 0.2943755 0.7492142 0.3396312 0.5404864 0.5499712 0.9230692
#> [32,] 0.6495209 0.5930319 0.8408753 0.5715279 0.6584678 0.7905140 0.9594931
#> [33,] 0.5565914 0.5937040 0.8379421 0.4734845 0.6106432 0.7699533 0.9624591
#> [34,] 0.5107974 0.5211500 0.7640057 0.3938363 0.5834255 0.7186390 0.9528629
#> [35,] 0.3996276 0.4796571 0.6627823 0.3554652 0.4404545 0.5644088 0.8905060
#> [36,] 0.4110733 0.5214083 0.7115310 0.3740066 0.4446186 0.6088137 0.9182536
#> [37,] 0.5399330 0.5274374 0.8204881 0.4869756 0.5813600 0.7444805 0.9577731
#> [38,] 0.5021416 0.3797547 0.7855157 0.4422491 0.6200093 0.6616724 0.9355965
#> [39,] 0.4195593 0.3985464 0.8015764 0.3921335 0.5159574 0.6177801 0.9449100
#> [40,] 0.6710397 0.6279731 0.8834268 0.5822620 0.6785677 0.8185025 0.9729130
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6568603 0.6759284 0.4044219
#>  [2,] 0.7535286 0.8175280 0.3972686
#>  [3,] 0.7922034 0.6664331 0.4440381
#>  [4,] 0.6210958 0.7648235 0.2700850
#>  [5,] 0.7539217 0.7876245 0.3056094
#>  [6,] 0.8624177 0.8241815 0.5328954
#>  [7,] 0.7303738 0.7309176 0.3365998
#>  [8,] 0.8156379 0.7664915 0.5177152
#>  [9,] 0.8004699 0.8378708 0.4617454
#> [10,] 0.7701209 0.7267218 0.4754743
#> [11,] 0.7158322 0.7786219 0.3108863
#> [12,] 0.7403104 0.8015640 0.3426108
#> [13,] 0.8410276 0.7872140 0.5133492
#> [14,] 0.7666560 0.6508059 0.4040867
#> [15,] 0.7900683 0.7337324 0.5338168
#> [16,] 0.7031438 0.6183549 0.3381523
#> [17,] 0.8386977 0.7949112 0.4483976
#> [18,] 0.7557933 0.6573925 0.4467487
#> [19,] 0.5817173 0.7072330 0.2696401
#> [20,] 0.7921102 0.7496651 0.4914165
#> [21,] 0.7611303 0.7853055 0.3763442
#> [22,] 0.7727603 0.7920890 0.3946373
#> [23,] 0.7231811 0.8040404 0.3085964
#> [24,] 0.7559471 0.7889368 0.3351052
#> [25,] 0.7969659 0.7927818 0.4502538
#> [26,] 0.8466055 0.7665335 0.4363826
#> [27,] 0.7436093 0.7555475 0.4234923
#> [28,] 0.7772955 0.7212004 0.4902421
#> [29,] 0.7396464 0.7954263 0.4250949
#> [30,] 0.7925362 0.8136268 0.4149208
#> [31,] 0.6708490 0.6683620 0.3135644
#> [32,] 0.8175041 0.7822235 0.5201541
#> [33,] 0.7928528 0.8147371 0.4081352
#> [34,] 0.7698783 0.7662583 0.3698962
#> [35,] 0.5578998 0.7323458 0.2529033
#> [36,] 0.6142393 0.7631736 0.2602389
#> [37,] 0.7762021 0.7805785 0.4234097
#> [38,] 0.7748696 0.6747149 0.4531579
#> [39,] 0.6875575 0.7478635 0.3237553
#> [40,] 0.8555580 0.8260048 0.5229932
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4922185 0.5077815
#>  [2,] 0.4816681 0.5183319
#>  [3,] 0.4952352 0.5047648
#>  [4,] 0.4983821 0.5016179
#>  [5,] 0.4928962 0.5071038
#>  [6,] 0.4744105 0.5255895
#>  [7,] 0.4990502 0.5009498
#>  [8,] 0.4772279 0.5227721
#>  [9,] 0.4747889 0.5252111
#> [10,] 0.4832291 0.5167709
#> [11,] 0.4921342 0.5078658
#> [12,] 0.4916279 0.5083721
#> [13,] 0.4809747 0.5190253
#> [14,] 0.5002670 0.4997330
#> [15,] 0.4766950 0.5233050
#> [16,] 0.5068096 0.4931904
#> [17,] 0.4844439 0.5155561
#> [18,] 0.4951927 0.5048073
#> [19,] 0.4963304 0.5036696
#> [20,] 0.4862097 0.5137903
#> [21,] 0.4839529 0.5160471
#> [22,] 0.4886016 0.5113984
#> [23,] 0.4908532 0.5091468
#> [24,] 0.4863917 0.5136083
#> [25,] 0.4801769 0.5198231
#> [26,] 0.4886650 0.5113350
#> [27,] 0.4838685 0.5161315
#> [28,] 0.4828602 0.5171398
#> [29,] 0.4865144 0.5134856
#> [30,] 0.4812633 0.5187367
#> [31,] 0.5072582 0.4927418
#> [32,] 0.4775601 0.5224399
#> [33,] 0.4830990 0.5169010
#> [34,] 0.4892455 0.5107545
#> [35,] 0.4966582 0.5033418
#> [36,] 0.4942967 0.5057033
#> [37,] 0.4872151 0.5127849
#> [38,] 0.4956665 0.5043335
#> [39,] 0.5004131 0.4995869
#> [40,] 0.4749331 0.5250669
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]        [,2]      [,3]         [,4]        [,5]      [,6]
#>  [1,]  0.025591686 -0.24859959 0.8564086 -0.252663372  0.25506101 0.4369568
#>  [2,]  0.392260047  0.38889049 1.4709676 -0.005025557  0.31197265 1.1075885
#>  [3,]  0.162069711 -0.49111630 1.3385945 -0.149546019  0.49095398 0.9275881
#>  [4,] -0.437686499 -0.17013492 1.0599921 -0.574249985  0.03868675 0.3202957
#>  [5,] -0.113655973  0.02343106 1.1104281 -0.509052899  0.21449336 0.7556082
#>  [6,]  0.764038552  0.52490370 2.0544086  0.381362626  0.81019347 1.5530884
#>  [7,] -0.081663871 -0.44438875 1.6389886 -0.361579686  0.32124435 0.7711817
#>  [8,]  0.703938054  0.36460102 1.7078609  0.398072328  0.66672160 1.3637870
#>  [9,]  0.600325752  0.69275038 1.8339693  0.292280815  0.62553600 1.3375623
#> [10,]  0.521677616  0.08509427 1.4222053  0.168019141  0.53162987 1.1133436
#> [11,] -0.245302832  0.12458344 1.2483864 -0.505134669  0.01422661 0.7663631
#> [12,] -0.090818887  0.05124084 1.4571342 -0.369023149  0.28511147 0.8239751
#> [13,]  0.537496320  0.19415012 1.8594008  0.179736334  0.73462647 1.2746346
#> [14,] -0.032185074 -0.69722052 1.2363047 -0.311434140  0.37929363 0.7063891
#> [15,]  0.540152864  0.40509110 1.3177882  0.244737684  0.71067575 1.1070567
#> [16,] -0.560587834 -0.82915087 1.0009193 -0.738913745  0.12626098 0.3269644
#> [17,]  0.305430028  0.17158412 1.7436224 -0.092681279  0.51919292 1.2361939
#> [18,]  0.118916337 -0.55586962 1.3386344 -0.277082736  0.46092118 0.5769542
#> [19,] -0.218862651 -0.17920238 0.6959260 -0.442820375 -0.05650833 0.2631328
#> [20,]  0.470528845 -0.10210445 1.6010151  0.169889390  0.62501762 1.0535121
#> [21,]  0.278687796  0.32624190 1.3854567 -0.060467245  0.31394069 1.0766376
#> [22,]  0.049558984  0.12378383 1.4713657 -0.115245263  0.25047575 1.0540166
#> [23,] -0.145712388  0.16894905 1.4230371 -0.391926019  0.14381983 0.7543657
#> [24,]  0.287624853  0.22682397 1.4259303 -0.102156279  0.36505094 1.0795983
#> [25,]  0.483266409  0.38444381 1.6398326  0.153069246  0.49388713 1.3399710
#> [26,]  0.326612196 -0.09845650 1.7319260 -0.096082923  0.66327526 1.2009688
#> [27,]  0.221600811  0.26260293 1.2283366 -0.041831220  0.35569948 0.9060980
#> [28,]  0.565554606  0.06055727 1.4659138  0.169820220  0.59757856 1.1217197
#> [29,]  0.270548310  0.07659422 1.5495204  0.017799152  0.34609684 0.9091875
#> [30,]  0.382375171  0.41859066 1.5395165  0.030455984  0.44477330 1.2760736
#> [31,] -0.476702061 -0.87422690 1.0944258 -0.664938352  0.16230080 0.2005545
#> [32,]  0.616933973  0.37651326 1.6647555  0.288087704  0.65647379 1.3280268
#> [33,]  0.227339890  0.37929904 1.6429953 -0.106161422  0.45001684 1.2080477
#> [34,]  0.043196142  0.08465042 1.1747674 -0.431214558  0.33685151 0.9377207
#> [35,] -0.407017006 -0.08141639 0.6757181 -0.595101481 -0.23931763 0.2590746
#> [36,] -0.359530265  0.08568542 0.9028315 -0.515066421 -0.22243805 0.4423285
#> [37,]  0.160072868  0.10985981 1.5196581 -0.052109363  0.32835880 1.0693877
#> [38,]  0.008566635 -0.49058948 1.2981041 -0.232039338  0.48958750 0.6707558
#> [39,] -0.324582861 -0.41152540 1.3961760 -0.438353098  0.06385135 0.4801366
#> [40,]  0.712891158  0.52353132 2.0252891  0.332065976  0.74719752 1.5062346
#>           [,7]      [,8]      [,9]       [,10]
#>  [1,] 2.161582 0.6493337 0.7351229 -0.38707369
#>  [2,] 3.207925 1.1175210 1.4996886 -0.41685900
#>  [3,] 3.025075 1.3382582 0.6920965 -0.22478924
#>  [4,] 2.377034 0.4942020 1.1793088 -0.99419160
#>  [5,] 3.095620 1.1196386 1.3106653 -0.82072697
#>  [6,] 3.646287 1.8355174 1.5449388  0.13177175
#>  [7,] 3.158816 0.9965198 0.9992832 -0.67848354
#>  [8,] 3.201899 1.4870685 1.1886054  0.07089039
#>  [9,] 3.352256 1.3892341 1.6424706 -0.15331790
#> [10,] 2.932143 1.2089938 0.9780535 -0.09818159
#> [11,] 2.740953 0.9238810 1.2576535 -0.79597883
#> [12,] 3.022471 1.0475826 1.3960985 -0.65168103
#> [13,] 3.377112 1.6658938 1.3082130  0.05340960
#> [14,] 2.913217 1.1895243 0.6225835 -0.38846586
#> [15,] 2.559193 1.3253369 1.0136427  0.13547393
#> [16,] 2.406917 0.8623134 0.4825713 -0.67153890
#> [17,] 3.360388 1.6485698 1.3547873 -0.20714716
#> [18,] 2.555469 1.1297531 0.6516959 -0.21381597
#> [19,] 2.193448 0.3298271 0.8819833 -0.99644938
#> [20,] 3.119161 1.3376921 1.0968270 -0.03433740
#> [21,] 3.122730 1.1588865 1.2968568 -0.50509442
#> [22,] 3.080243 1.2239636 1.3375639 -0.42786061
#> [23,] 2.894919 0.9602964 1.4117411 -0.80668939
#> [24,] 3.285764 1.1305865 1.3185286 -0.68518454
#> [25,] 3.281210 1.3674379 1.3417754 -0.19964543
#> [26,] 3.487112 1.7082221 1.1888400 -0.25585615
#> [27,] 2.729627 1.0648133 1.1284214 -0.30845318
#> [28,] 2.934122 1.2499748 0.9504240 -0.03903639
#> [29,] 3.063525 1.0441313 1.3579497 -0.30189258
#> [30,] 3.342757 1.3402814 1.4737510 -0.34365969
#> [31,] 2.484798 0.7120273 0.7007858 -0.78350777
#> [32,] 3.164932 1.4995282 1.2786712  0.08066030
#> [33,] 3.244060 1.3422081 1.4810898 -0.37167987
#> [34,] 3.006411 1.2076242 1.1873028 -0.53266217
#> [35,] 2.095920 0.2326430 1.0065568 -1.08318741
#> [36,] 2.418851 0.4651673 1.1701582 -1.04472702
#> [37,] 3.121554 1.2436695 1.2690405 -0.30879168
#> [38,] 2.676017 1.2360150 0.7295881 -0.18791935
#> [39,] 2.842121 0.7887249 1.0872497 -0.73656689
#> [40,] 3.581239 1.7788755 1.5575726  0.09203760
#> 
#> $pre[[3]]
#>              [,1]          [,2]
#>  [1,] -0.05303709 -0.0219086464
#>  [2,] -0.06822186  0.0051385218
#>  [3,] -0.03378110 -0.0147211618
#>  [4,] -0.03955314 -0.0330813322
#>  [5,] -0.04165308 -0.0132357885
#>  [6,] -0.07773509  0.0247125452
#>  [7,] -0.02934309 -0.0255437730
#>  [8,] -0.07328946  0.0178621092
#>  [9,] -0.07978764  0.0211422678
#> [10,] -0.06341054  0.0036983354
#> [11,] -0.04438882 -0.0129231521
#> [12,] -0.04591773 -0.0124263159
#> [13,] -0.06493423  0.0112037777
#> [14,] -0.02419984 -0.0252677937
#> [15,] -0.07713896  0.0161487444
#> [16,] -0.01266166 -0.0399018796
#> [17,] -0.05585696  0.0063874095
#> [18,] -0.03980750 -0.0205776380
#> [19,] -0.04677644 -0.0320976240
#> [20,] -0.05695771 -0.0017824230
#> [21,] -0.06025984  0.0039505991
#> [22,] -0.04890711 -0.0033057133
#> [23,] -0.04750815 -0.0109167161
#> [24,] -0.05491333 -0.0004666653
#> [25,] -0.06710593  0.0122281367
#> [26,] -0.04587528 -0.0005274149
#> [27,] -0.06293076  0.0016177375
#> [28,] -0.06438287  0.0042030306
#> [29,] -0.05955393 -0.0055982805
#> [30,] -0.06499076  0.0099909917
#> [31,] -0.01684341 -0.0458783154
#> [32,] -0.07341915  0.0164009082
#> [33,] -0.06080658  0.0068229949
#> [34,] -0.04947068 -0.0064459204
#> [35,] -0.04751316 -0.0341459079
#> [36,] -0.04701197 -0.0241977828
#> [37,] -0.05273150 -0.0015807474
#> [38,] -0.03572164 -0.0183873449
#> [39,] -0.02996543 -0.0316178068
#> [40,] -0.07701633  0.0233354982
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5077815 -0.5077815
#>  [2,]  0.5183319 -0.5183319
#>  [3,]  0.5047648 -0.5047648
#>  [4,]  0.5016179 -0.5016179
#>  [5,] -0.4928962  0.4928962
#>  [6,] -0.4744105  0.4744105
#>  [7,]  0.5009498 -0.5009498
#>  [8,]  0.5227721 -0.5227721
#>  [9,] -0.4747889  0.4747889
#> [10,]  0.5167709 -0.5167709
#> [11,]  0.5078658 -0.5078658
#> [12,]  0.5083721 -0.5083721
#> [13,] -0.4809747  0.4809747
#> [14,]  0.4997330 -0.4997330
#> [15,] -0.4766950  0.4766950
#> [16,] -0.5068096  0.5068096
#> [17,]  0.5155561 -0.5155561
#> [18,] -0.4951927  0.4951927
#> [19,]  0.5036696 -0.5036696
#> [20,] -0.4862097  0.4862097
#> [21,]  0.5160471 -0.5160471
#> [22,]  0.5113984 -0.5113984
#> [23,]  0.5091468 -0.5091468
#> [24,] -0.4863917  0.4863917
#> [25,]  0.5198231 -0.5198231
#> [26,] -0.4886650  0.4886650
#> [27,]  0.5161315 -0.5161315
#> [28,]  0.5171398 -0.5171398
#> [29,]  0.5134856 -0.5134856
#> [30,]  0.5187367 -0.5187367
#> [31,] -0.5072582  0.5072582
#> [32,] -0.4775601  0.4775601
#> [33,]  0.5169010 -0.5169010
#> [34,] -0.4892455  0.4892455
#> [35,]  0.5033418 -0.5033418
#> [36,] -0.4942967  0.4942967
#> [37,]  0.5127849 -0.5127849
#> [38,] -0.4956665  0.4956665
#> [39,] -0.5004131  0.5004131
#> [40,] -0.4749331  0.4749331
#> 
#> $L
#> [1] 0.6927954 0.6921117 0.6924588 0.6895659 0.6899391 0.6968323
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4492754