Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1           V10          V11         V12          V13
#>  [1,] -0.1960433 -0.0937471344 -0.082053233 -0.08248932  0.006782529
#>  [2,] -0.4117328 -0.0008859411 -0.060918987  0.04217654  0.002912372
#>  [3,] -0.3628492 -0.0994081740 -0.119672678 -0.04283834 -0.182103353
#>  [4,] -0.2744482 -0.0349361646 -0.144338631 -0.11428957 -0.093756302
#>  [5,] -0.3343153 -0.1497954122 -0.005156061 -0.07477119 -0.141760600
#>  [6,] -0.2723621 -0.1383990359 -0.161976764 -0.15120414 -0.110402759
#>  [7,] -0.2108330 -0.0216055790 -0.044099956 -0.04672368 -0.146913282
#>  [8,] -0.3167068 -0.1015639304 -0.104088253 -0.16187249  0.036750650
#>  [9,] -0.3807224 -0.1639626700 -0.098264353 -0.10924277 -0.138907409
#> [10,] -0.2120491 -0.1358576104 -0.127383407 -0.04696936 -0.068913761
#>                V14         V15          V16          V17         V18
#>  [1,] -0.031108830  0.09413892  0.007964718  0.026527909  0.12110587
#>  [2,]  0.049044726  0.02743204  0.092535463  0.053306428  0.01289690
#>  [3,] -0.084192646 -0.03487147 -0.130740859 -0.077201857 -0.02626631
#>  [4,] -0.131076961 -0.04112030 -0.122980718 -0.096705886 -0.04526754
#>  [5,]  0.001380209  0.01854675 -0.017234349 -0.096744419  0.05021931
#>  [6,] -0.039653968 -0.12615186  0.034053464 -0.006694311 -0.03985609
#>  [7,] -0.089118078 -0.14622751  0.060697571  0.085631481  0.08269690
#>  [8,] -0.024299677  0.03298757  0.038552211  0.056618333  0.12484425
#>  [9,] -0.061905716 -0.12996023 -0.073872152  0.021435018  0.11948961
#> [10,] -0.169990782 -0.03903915 -0.168316574 -0.071091193 -0.10119456
#>               V19         V2         V20          V21         V22         V23
#>  [1,]  0.10330448 -0.2982500  0.21213399  0.186171876 0.166373091 0.279628959
#>  [2,]  0.11832941 -0.2941530  0.09873519  0.266401052 0.294529040 0.266731059
#>  [3,] -0.02737274 -0.2509412 -0.01119222  0.084583158 0.078003538 0.159446204
#>  [4,]  0.06140262 -0.2386252 -0.01282269 -0.001771125 0.095693366 0.089138284
#>  [5,] -0.03637862 -0.2490118  0.05119455  0.106814250 0.137665298 0.002761109
#>  [6,]  0.13051556 -0.2165188  0.10270886  0.043760669 0.075330990 0.077134037
#>  [7,]  0.03387247 -0.2810344  0.15973523  0.057576045 0.067927087 0.107254836
#>  [8,]  0.15639386 -0.2247105  0.14562427  0.135900014 0.087779913 0.073723572
#>  [9,]  0.14873674 -0.3105238  0.11278889  0.035706193 0.206302851 0.149314726
#> [10,]  0.01358973 -0.3455248  0.06110432  0.093138554 0.005559275 0.070639620
#>              V24        V25        V26         V27        V28        V29
#>  [1,] 0.14708821 0.20909711 0.25867946 0.311462757 0.29519395 0.16750669
#>  [2,] 0.14873854 0.14815929 0.16180577 0.194242958 0.26515126 0.07076376
#>  [3,] 0.10977456 0.02432799 0.04631151 0.081141511 0.07387700 0.06731284
#>  [4,] 0.02914627 0.09368876 0.07639557 0.235379974 0.22546470 0.20521507
#>  [5,] 0.18826865 0.14660454 0.09142008 0.171880703 0.14784600 0.01003966
#>  [6,] 0.22713323 0.07227294 0.26506474 0.110792863 0.06387642 0.04304087
#>  [7,] 0.19793229 0.21070977 0.12468199 0.070966627 0.15235820 0.02351001
#>  [8,] 0.06933049 0.11725659 0.10640244 0.001035939 0.03777682 0.09391760
#>  [9,] 0.05968379 0.26176381 0.15035191 0.075820196 0.19808375 0.12525819
#> [10,] 0.22124960 0.22510006 0.14952244 0.245974862 0.10677918 0.04151235
#>               V3          V30         V31         V32          V33          V34
#>  [1,] -0.2445287  0.143014251  0.05563712  0.06395550 -0.018783362  0.009889703
#>  [2,] -0.1952795  0.103226184  0.14476597 -0.04477016  0.043415332  0.088668744
#>  [3,] -0.2203228  0.087433804 -0.06945443  0.05281134  0.038234607 -0.086573440
#>  [4,] -0.2758461  0.088296617  0.15572488 -0.04516375 -0.076289148  0.086776828
#>  [5,] -0.2633797 -0.048041073 -0.06431694 -0.09765712  0.043305546 -0.070714753
#>  [6,] -0.3063848  0.123106962 -0.02180558 -0.12453488 -0.012508852 -0.147678277
#>  [7,] -0.1897108  0.058187814 -0.01065076  0.03767933 -0.002981297 -0.104397474
#>  [8,] -0.3271108  0.083015344 -0.07885181 -0.11437193  0.088228371  0.073836453
#>  [9,] -0.3613397  0.018275136  0.07895076  0.01378122 -0.086282620 -0.044400082
#> [10,] -0.2597014 -0.001443269  0.07017409  0.05460245 -0.117181561  0.007598161
#>                V35         V36           V37          V38          V39
#>  [1,]  0.108960823 -0.03256188 -4.453289e-02  0.010111930 -0.052347065
#>  [2,]  0.015313409  0.03310175  6.271740e-02  0.034144267 -0.039475082
#>  [3,] -0.006047959 -0.02017800 -1.102150e-01 -0.089253643 -0.025101845
#>  [4,]  0.040765286  0.06477548 -4.846397e-02 -0.062592804 -0.158563863
#>  [5,] -0.097634341 -0.02544765 -5.683635e-02 -0.130946791 -0.092852610
#>  [6,] -0.135598891 -0.07034996 -1.793307e-01 -0.118524628 -0.121524684
#>  [7,] -0.033658334 -0.15052164 -1.001897e-01 -0.091741573  0.002171888
#>  [8,] -0.087758914 -0.07180886  6.856639e-02  0.001744622 -0.025075962
#>  [9,] -0.046664146 -0.09584381 -8.919545e-02 -0.062660843 -0.079605611
#> [10,] -0.140856120  0.02658006 -1.070931e-07 -0.103370247 -0.005603193
#>               V4          V40         V41          V42         V43         V44
#>  [1,] -0.2897920  0.043103508 -0.05478765 -0.028969840 -0.03336903 -0.10511418
#>  [2,] -0.2815096 -0.006500847 -0.08430714 -0.104877272 -0.01886133 -0.04488171
#>  [3,] -0.2987135 -0.016121909 -0.19521688 -0.162660823 -0.14806606 -0.06905662
#>  [4,] -0.1477546 -0.063667251 -0.07674852 -0.126146399 -0.13648322 -0.17534444
#>  [5,] -0.1527685 -0.010024804  0.02905665 -0.112982644 -0.08218878 -0.07261556
#>  [6,] -0.2230718 -0.180278590 -0.14789162 -0.003165303 -0.01439978 -0.10157552
#>  [7,] -0.3127370 -0.005860831 -0.12948216 -0.126934378 -0.10771460 -0.13135089
#>  [8,] -0.3427365 -0.143558395 -0.10862050 -0.046926765 -0.13406420 -0.14892308
#>  [9,] -0.2683056 -0.012974446 -0.02859897 -0.039302570  0.02541454 -0.14713305
#> [10,] -0.1779349 -0.159931736 -0.10466617 -0.067532383 -0.17847853 -0.13087991
#>               V45         V46         V47        V48        V49          V5
#>  [1,] -0.14277857 -0.07646608 -0.17584135 -0.2421690 -0.2370420 -0.07566585
#>  [2,] -0.09008394 -0.11192538 -0.21144207 -0.2512221 -0.2817484 -0.24129100
#>  [3,] -0.09187416 -0.11969156 -0.11520426 -0.2262037 -0.1774238 -0.23740919
#>  [4,] -0.19666240 -0.16821915 -0.06331927 -0.2734005 -0.1702128 -0.15147235
#>  [5,] -0.13546528 -0.15193040 -0.26708465 -0.2605764 -0.3118076 -0.30024185
#>  [6,] -0.08291362 -0.10078884 -0.14392489 -0.2339881 -0.3595226 -0.29444031
#>  [7,] -0.14869966 -0.10071545 -0.20762894 -0.2058690 -0.2750653 -0.24948250
#>  [8,] -0.06591952 -0.19843547 -0.24307742 -0.1924565 -0.2705289 -0.30091228
#>  [9,] -0.14585610 -0.17173885 -0.11529436 -0.1086505 -0.3198620 -0.19294876
#> [10,] -0.18233262 -0.17700107 -0.18349876 -0.1374558 -0.2554524 -0.25940684
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2268825 -0.2499073 -0.3316639 -0.2802231 -0.2472393 -0.2198869
#>  [2,] -0.2685626 -0.3115489 -0.3980026 -0.3883626 -0.4107460 -0.3818659
#>  [3,] -0.3127406 -0.2483749 -0.2067749 -0.3563145 -0.3237000 -0.2404794
#>  [4,] -0.2429537 -0.2847213 -0.3084210 -0.2140233 -0.2100744 -0.3014070
#>  [5,] -0.2314676 -0.3091461 -0.3435232 -0.4030433 -0.3400347 -0.2364921
#>  [6,] -0.2804101 -0.3715393 -0.3306885 -0.2377480 -0.3161920 -0.4243266
#>  [7,] -0.3094943 -0.3622073 -0.3508564 -0.2612482 -0.2288651 -0.4232602
#>  [8,] -0.3256154 -0.3894337 -0.3863064 -0.3013030 -0.2814086 -0.3245003
#>  [9,] -0.3888337 -0.3664490 -0.3699215 -0.4065553 -0.3649379 -0.3458233
#> [10,] -0.3668185 -0.2069050 -0.2541146 -0.3291632 -0.3652909 -0.3061584
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.2917052 -0.2767852 -0.1924778 -0.2650826 -0.15609247 -0.2418733
#>  [2,] -0.3451845 -0.2820028 -0.4083898 -0.4346188 -0.21550784 -0.3238566
#>  [3,] -0.3074814 -0.3369405 -0.3147907 -0.2327184 -0.21497887 -0.2462400
#>  [4,] -0.3560190 -0.4046385 -0.2690300 -0.2587857 -0.23304950 -0.3995695
#>  [5,] -0.3332302 -0.4038209 -0.2698084 -0.2648318 -0.20903376 -0.3481858
#>  [6,] -0.4026617 -0.2960680 -0.3888732 -0.4378535 -0.21771451 -0.4186502
#>  [7,] -0.3650351 -0.2661962 -0.3878047 -0.2778658 -0.11228377 -0.3210409
#>  [8,] -0.2992411 -0.3093197 -0.2456680 -0.2860361 -0.25646947 -0.2725083
#>  [9,] -0.2570450 -0.3367831 -0.3989840 -0.3812782 -0.08987393 -0.4342558
#> [10,] -0.2898386 -0.3394348 -0.3906351 -0.3104920 -0.24230583 -0.3162587
#>                V7          V8          V9
#>  [1,] -0.17604762 -0.17494067 -0.09384765
#>  [2,] -0.20262539 -0.10960724 -0.11951877
#>  [3,] -0.20662153 -0.09180520 -0.14031925
#>  [4,] -0.19228810 -0.22365821 -0.15428183
#>  [5,] -0.09807509 -0.20679206 -0.12539362
#>  [6,] -0.25463936 -0.22633828 -0.14107119
#>  [7,] -0.22791082 -0.07166165 -0.09964481
#>  [8,] -0.10908722 -0.11298609 -0.07830922
#>  [9,] -0.07821967 -0.16818970 -0.15159543
#> [10,] -0.25332453 -0.25716344 -0.17371195
#> 
#> $W[[2]]
#>            [,1]       [,2]        [,3]        [,4]       [,5]        [,6]
#> [1,] 0.10500658 0.01903286  0.04197472 -0.12243431 0.05875281  0.06899245
#> [2,] 0.05292989 0.01230704 -0.04099930 -0.01671435 0.02357793 -0.01860750
#>             [,7]         [,8]         [,9]       [,10]
#> [1,]  0.05376676 -0.067513398  0.007701383 -0.04845235
#> [2,] -0.04498712 -0.006667639 -0.079877335  0.12194833
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  3.467920e-05  1.869439e-04  0.0002068030  0.0002347655  2.845987e-04
#>  [2,]  9.811879e-05  4.721702e-04  0.0005488323  0.0005637687  5.772848e-04
#>  [3,] -1.405201e-04 -6.757783e-04 -0.0009268537 -0.0008959557 -6.165096e-04
#>  [4,]  2.415488e-04  1.122810e-03  0.0013633127  0.0014255326  1.360357e-03
#>  [5,] -2.040777e-05 -7.636814e-05 -0.0001492276 -0.0001019430  4.410858e-05
#>  [6,] -1.265664e-04 -5.655365e-04 -0.0007904936 -0.0007629953 -5.194878e-04
#>  [7,] -1.187682e-04 -5.230309e-04 -0.0006953175 -0.0006381462 -4.237256e-04
#>  [8,]  1.885578e-04  9.267588e-04  0.0011157327  0.0011566389  1.110119e-03
#>  [9,] -7.776018e-05 -3.198886e-04 -0.0004302592 -0.0003650363 -2.023142e-04
#> [10,]  3.627491e-04  1.686750e-03  0.0021064100  0.0021834387  1.991192e-03
#>                 V14           V15          V16          V17          V18
#>  [1,]  2.913226e-04  3.171771e-04 0.0003680588 0.0003865652 0.0003809537
#>  [2,]  5.400401e-04  5.462364e-04 0.0005644563 0.0005143310 0.0004998359
#>  [3,] -2.914362e-04 -7.905055e-05 0.0001639716 0.0003797146 0.0003240395
#>  [4,]  1.119414e-03  1.002172e-03 0.0008971063 0.0006430867 0.0006648410
#>  [5,]  1.542964e-04  2.385084e-04 0.0003551768 0.0004314485 0.0004170043
#>  [6,] -2.408347e-04 -5.241470e-05 0.0001732802 0.0003712534 0.0003361172
#>  [7,] -2.069555e-04 -6.234670e-05 0.0001349006 0.0003458289 0.0003318191
#>  [8,]  9.312040e-04  8.472875e-04 0.0007792506 0.0006041680 0.0006212783
#>  [9,] -6.887984e-05  2.233122e-05 0.0001698608 0.0003253920 0.0003171109
#> [10,]  1.537619e-03  1.278281e-03 0.0010231038 0.0006124223 0.0006550833
#>                 V19            V2           V20           V21           V22
#>  [1,]  4.031295e-04  4.255048e-05  0.0004358768  0.0004737147  0.0004302666
#>  [2,]  7.015289e-04  1.191487e-04  0.0008783162  0.0009511706  0.0007733042
#>  [3,] -3.903212e-04 -1.647248e-04 -0.0008505193 -0.0008972076 -0.0005249625
#>  [4,]  1.459149e-03  2.876865e-04  0.0021184689  0.0023236150  0.0017890675
#>  [5,]  2.152923e-04 -2.148297e-05  0.0001097648  0.0001317306  0.0002223570
#>  [6,] -2.846777e-04 -1.467461e-04 -0.0006875701 -0.0007341201 -0.0003988106
#>  [7,] -2.113960e-04 -1.362582e-04 -0.0005959778 -0.0006416466 -0.0003109921
#>  [8,]  1.204519e-03  2.261206e-04  0.0016812751  0.0018304528  0.0014357625
#>  [9,] -4.678653e-05 -8.739931e-05 -0.0003092553 -0.0003328494 -0.0001013376
#> [10,]  1.950913e-03  4.313038e-04  0.0029606561  0.0032418095  0.0024354673
#>                 V23           V24          V25          V26          V27
#>  [1,]  4.479500e-04  4.751123e-04 0.0004880969 0.0005177648 0.0005724955
#>  [2,]  7.164446e-04  7.387417e-04 0.0006876497 0.0007159951 0.0008459048
#>  [3,] -1.714128e-04 -1.468284e-05 0.0003097706 0.0003766749 0.0002145038
#>  [4,]  1.609154e-03  1.606721e-03 0.0013470541 0.0013904539 0.0017323560
#>  [5,]  3.831673e-04  4.640470e-04 0.0005774794 0.0006253122 0.0006319159
#>  [6,] -9.679613e-05  2.970850e-05 0.0003065161 0.0003643024 0.0002225272
#>  [7,] -3.962982e-05  5.556458e-05 0.0002857177 0.0003261428 0.0002066430
#>  [8,]  1.303143e-03  1.322825e-03 0.0011610691 0.0012210171 0.0014873745
#>  [9,]  1.127823e-04  1.818120e-04 0.0003291383 0.0003666604 0.0003190506
#> [10,]  2.095565e-03  2.040034e-03 0.0015912270 0.0016235028 0.0021397718
#>                 V28           V29            V3          V30          V31
#>  [1,]  6.262740e-04  5.692283e-04  4.713987e-05 4.860178e-04 0.0004500436
#>  [2,]  1.036027e-03  9.535625e-04  1.312608e-04 8.064266e-04 0.0006302550
#>  [3,] -1.218263e-04 -9.448575e-05 -1.817946e-04 1.054984e-04 0.0005422762
#>  [4,]  2.262160e-03  2.115705e-03  2.957831e-04 1.649573e-03 0.0010552348
#>  [5,]  5.729741e-04  5.353940e-04 -2.664088e-05 5.075604e-04 0.0006169348
#>  [6,] -6.258834e-05 -3.997247e-05 -1.630260e-04 1.286712e-04 0.0005159769
#>  [7,] -7.590143e-05 -6.916542e-05 -1.461667e-04 5.340578e-05 0.0004240840
#>  [8,]  1.882387e-03  1.741615e-03  2.397572e-04 1.374365e-03 0.0009234013
#>  [9,]  1.356463e-04  1.175320e-04 -9.214836e-05 1.545971e-04 0.0004080983
#> [10,]  2.913412e-03  2.672521e-03  4.551838e-04 1.970719e-03 0.0010647261
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0004136741 3.960105e-04 0.0003799952 0.0003892351 3.514762e-04
#>  [2,] 0.0006583065 6.469265e-04 0.0005815850 0.0005375688 3.725765e-04
#>  [3,] 0.0001703597 1.361479e-04 0.0002978835 0.0005316437 9.118370e-04
#>  [4,] 0.0012819077 1.329776e-03 0.0011492539 0.0008826338 4.439785e-04
#>  [5,] 0.0004517890 4.183853e-04 0.0004554178 0.0005216036 6.321316e-04
#>  [6,] 0.0001774116 1.351065e-04 0.0002780280 0.0004988576 8.473225e-04
#>  [7,] 0.0001288596 6.406473e-05 0.0001797308 0.0003609706 6.671289e-04
#>  [8,] 0.0010633300 1.057196e-03 0.0009028855 0.0007102537 3.662596e-04
#>  [9,] 0.0001905121 1.278819e-04 0.0001977898 0.0003155902 5.353677e-04
#> [10,] 0.0015162919 1.551423e-03 0.0012290936 0.0007878659 6.582767e-05
#>                V37           V38           V39            V4           V40
#>  [1,] 0.0002919500  3.055494e-04  2.856699e-04  8.268691e-05  2.906405e-04
#>  [2,] 0.0003548796  5.636577e-04  6.121034e-04  2.358442e-04  5.149921e-04
#>  [3,] 0.0005653823 -1.641567e-04 -5.208476e-04 -3.330054e-04 -1.032642e-04
#>  [4,] 0.0004370968  1.178001e-03  1.458285e-03  5.745032e-04  1.102886e-03
#>  [5,] 0.0004360141  2.015291e-04  7.267346e-05 -4.524315e-05  2.407728e-04
#>  [6,] 0.0005429592 -1.291567e-04 -4.577520e-04 -3.031466e-04 -7.982689e-05
#>  [7,] 0.0004121654 -1.767070e-04 -4.492708e-04 -2.806241e-04 -1.026520e-04
#>  [8,] 0.0003796714  9.294926e-04  1.134311e-03  4.498743e-04  8.870066e-04
#>  [9,] 0.0003430408 -6.349445e-05 -2.475122e-04 -1.779440e-04  1.215749e-05
#> [10,] 0.0002089668  1.483423e-03  2.024952e-03  8.761068e-04  1.430594e-03
#>                 V41           V42           V43           V44           V45
#>  [1,]  2.891757e-04  0.0002790132  2.392725e-04  2.274225e-04  0.0001983625
#>  [2,]  5.302762e-04  0.0006082611  5.592703e-04  5.759661e-04  0.0006683228
#>  [3,] -2.235200e-04 -0.0006410790 -6.975519e-04 -8.226610e-04 -0.0012701257
#>  [4,]  1.121103e-03  0.0015139892  1.480063e-03  1.554458e-03  0.0018449998
#>  [5,]  1.851424e-04  0.0000390344 -6.968847e-06 -5.314708e-05 -0.0002521073
#>  [6,] -1.904959e-04 -0.0005754613 -6.291590e-04 -7.344822e-04 -0.0011471154
#>  [7,] -1.791738e-04 -0.0005128385 -5.637756e-04 -6.639742e-04 -0.0010547175
#>  [8,]  9.318991e-04  0.0012148577  1.167179e-03  1.248847e-03  0.0014858707
#>  [9,] -3.399046e-05 -0.0002666246 -3.157210e-04 -3.855153e-04 -0.0006801148
#> [10,]  1.534865e-03  0.0022201296  2.199688e-03  2.349462e-03  0.0029035048
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0001504497  0.0001061946  8.333166e-05  4.551548e-05  1.030525e-04
#>  [2,]  0.0005503294  0.0003417568  2.505834e-04  1.358485e-04  2.546314e-04
#>  [3,] -0.0011111839 -0.0006114054 -4.070771e-04 -2.265572e-04 -3.353692e-04
#>  [4,]  0.0015235146  0.0009206043  6.708829e-04  3.716300e-04  6.260529e-04
#>  [5,] -0.0002476982 -0.0001050928 -5.628652e-05 -3.251310e-05 -2.388864e-05
#>  [6,] -0.0010009708 -0.0005479003 -3.671077e-04 -2.051355e-04 -2.966462e-04
#>  [7,] -0.0009205500 -0.0004954965 -3.329983e-04 -1.860659e-04 -2.641281e-04
#>  [8,]  0.0012227755  0.0007489018  5.406802e-04  2.964421e-04  4.914285e-04
#>  [9,] -0.0006005294 -0.0003086253 -2.050051e-04 -1.141648e-04 -1.531728e-04
#> [10,]  0.0024185481  0.0014448020  1.035245e-03  5.749109e-04  9.430571e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.115301e-05  1.576135e-05  1.360019e-05  1.096780e-05  1.219390e-05
#>  [2,]  5.106920e-05  4.040113e-05  4.106335e-05  2.465475e-05  2.668692e-05
#>  [3,] -5.287855e-05 -5.868529e-05 -6.518557e-05 -2.359004e-05 -2.816561e-05
#>  [4,]  1.265431e-04  1.045228e-04  1.092993e-04  5.816599e-05  6.907750e-05
#>  [5,]  1.948473e-06 -5.505994e-06 -1.009780e-05  1.440996e-06  1.499851e-06
#>  [6,] -4.919715e-05 -5.144674e-05 -5.932639e-05 -2.125619e-05 -2.521997e-05
#>  [7,] -4.715389e-05 -4.572797e-05 -5.462465e-05 -1.969415e-05 -2.311172e-05
#>  [8,]  9.994746e-05  8.366652e-05  8.405267e-05  4.571918e-05  5.245753e-05
#>  [9,] -2.816617e-05 -2.721955e-05 -3.489150e-05 -1.147809e-05 -1.328092e-05
#> [10,]  1.835944e-04  1.574980e-04  1.665760e-04  8.304173e-05  9.815355e-05
#>                 V55           V56          V57           V58           V59
#>  [1,]  9.946675e-06  6.066683e-06 7.006191e-06  6.095419e-06  9.016885e-06
#>  [2,]  2.222543e-05  1.231024e-05 1.212765e-05  1.481922e-05  2.200075e-05
#>  [3,] -1.837036e-05 -6.291642e-06 3.959301e-06 -1.617754e-05 -2.487603e-05
#>  [4,]  5.148047e-05  2.394452e-05 2.117375e-05  3.183342e-05  5.152238e-05
#>  [5,]  2.087475e-06  3.064316e-06 7.783467e-06 -1.209840e-06 -1.185843e-06
#>  [6,] -1.714882e-05 -5.206588e-06 3.899174e-06 -1.415814e-05 -2.267796e-05
#>  [7,] -1.694396e-05 -5.074814e-06 2.124841e-06 -1.303863e-05 -2.084835e-05
#>  [8,]  4.005737e-05  2.102120e-05 1.741465e-05  2.515138e-05  4.020208e-05
#>  [9,] -1.033983e-05 -2.502954e-06 2.805138e-06 -8.399860e-06 -1.267199e-05
#> [10,]  7.185314e-05  3.302504e-05 2.413963e-05  4.652925e-05  7.590171e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.059217e-04  8.568950e-06  1.364371e-04  1.265758e-04  1.741523e-04
#>  [2,]  1.877519e-04  2.034034e-05  2.558914e-04  2.682879e-04  4.336615e-04
#>  [3,] -9.870203e-05 -1.829456e-05 -1.379582e-04 -2.649003e-04 -5.967699e-04
#>  [4,]  3.904579e-04  4.214416e-05  5.654120e-04  6.242946e-04  1.048024e-03
#>  [5,]  6.190603e-05  5.701854e-08  8.163356e-05  2.574449e-05 -5.227253e-05
#>  [6,] -7.003263e-05 -1.648836e-05 -1.080972e-04 -2.104135e-04 -5.039860e-04
#>  [7,] -5.420926e-05 -1.639297e-05 -1.015636e-04 -1.980993e-04 -4.667911e-04
#>  [8,]  3.260932e-04  3.437238e-05  4.597544e-04  5.135519e-04  8.654264e-04
#>  [9,] -4.717296e-06 -1.038685e-05 -3.352077e-05 -1.046840e-04 -2.817510e-04
#> [10,]  5.338254e-04  6.074127e-05  7.700155e-04  8.926468e-04  1.570883e-03
#> 
#> $vW[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#> [1,] -0.03412295 -0.03383261 -0.008458118 -0.01038802 -0.01311633 -0.01141408
#> [2,]  0.03412295  0.03383261  0.008458118  0.01038802  0.01311633  0.01141408
#>             [,7]        [,8]        [,9]        [,10]
#> [1,] -0.01746794 -0.01556201 -0.02154208 -0.009434107
#> [2,]  0.01746794  0.01556201  0.02154208  0.009434107
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5348032 0.5811956 0.4014081 0.4547065 0.5406442 0.5046247 0.5817124
#>  [8] 0.4524709 0.7802273 0.4715802
#> 
#> $B[[2]]
#> [1] -0.05477121 -0.13437379
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0009135043 0.0014416959 0.0001930038 0.0029739567 0.0009298870
#>  [6] 0.0002049614 0.0001452455 0.0024080640 0.0003312590 0.0036408159
#> 
#> $vB[[2]]
#> [1] -0.04844827  0.04844827
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0209 0.1853 0.1748 0.1556 0.1476 0.1378 0.2584 0.3827 0.4784 0.5360
#>  [2,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#>  [3,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#>  [4,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#>  [5,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#>  [6,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [7,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#>  [8,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [9,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [10,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [11,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [12,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [13,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [14,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [15,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [16,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [17,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [18,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [19,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [20,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [21,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [22,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [23,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [24,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [25,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [26,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [27,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [28,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [29,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [30,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [31,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [32,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [33,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [34,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [35,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [36,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [37,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [38,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [39,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [40,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6192 0.0278 0.7912 0.9264 1.0000 0.9080 0.7435 0.5557 0.3172 0.1295
#>  [2,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#>  [3,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#>  [4,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#>  [5,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#>  [6,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [7,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#>  [8,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [9,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [10,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [11,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [12,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [13,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [14,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [15,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [16,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [17,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [18,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [19,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [20,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [21,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [22,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [23,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [24,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [25,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [26,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [27,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [28,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [29,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [30,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [31,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [32,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [33,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [34,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [35,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [36,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [37,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [38,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [39,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [40,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.0598 0.2722 0.0115 0.3616 0.3293 0.4855 0.3936 0.1845 0.0342 0.2489
#>  [2,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#>  [3,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#>  [4,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#>  [5,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#>  [6,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [7,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#>  [8,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [9,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [10,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [11,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [12,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [13,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [14,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [15,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [16,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [17,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [18,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [19,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [20,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [21,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [22,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [23,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [24,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [25,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [26,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [27,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [28,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [29,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [30,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [31,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [32,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [33,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [34,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [35,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [36,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [37,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [38,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [39,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [40,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3837 0.3514 0.2654 0.0445 0.1760 0.1599 0.0866 0.0590 0.0813 0.0492
#>  [2,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#>  [3,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#>  [4,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#>  [5,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#>  [6,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [7,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#>  [8,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [9,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [10,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [11,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [12,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [13,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [14,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [15,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [16,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [17,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [18,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [19,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [20,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [21,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [22,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [23,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [24,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [25,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [26,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [27,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [28,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [29,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [30,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [31,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [32,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [33,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [34,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [35,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [36,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [37,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [38,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [39,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [40,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0417 0.0495 0.0367 0.0115 0.0427 0.0118 0.0133 0.0096 0.0014 0.0049
#>  [2,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#>  [3,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#>  [4,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#>  [5,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#>  [6,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [7,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#>  [8,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [9,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [10,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [11,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [12,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [13,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [14,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [15,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [16,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [17,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [18,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [19,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [20,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [21,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [22,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [23,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [24,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [25,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [26,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [27,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [28,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [29,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [30,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [31,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [32,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [33,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [34,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [35,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [36,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [37,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [38,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [39,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [40,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0039 0.0029 0.0078 0.0047 0.0021 0.0766 0.0011 0.1458 0.1430 0.1894
#>  [2,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#>  [3,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#>  [4,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#>  [5,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#>  [6,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [7,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#>  [8,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [9,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [10,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [11,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [12,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [13,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [14,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [15,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [16,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [17,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [18,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [19,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [20,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [21,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [22,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [23,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [24,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [25,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [26,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [27,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [28,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [29,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [30,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [31,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [32,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [33,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [34,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [35,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [36,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [37,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [38,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [39,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [40,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8487213 0.8619279 0.5838326 0.6002492 0.6510815 0.6331279 0.7283874
#>  [2,] 0.8392108 0.8226875 0.4731905 0.5878197 0.5212794 0.4865927 0.5832983
#>  [3,] 0.8834163 0.8751612 0.5935260 0.7331052 0.6410215 0.6219356 0.6995615
#>  [4,] 0.8005196 0.8114798 0.4119759 0.4490061 0.5389468 0.5048367 0.6156627
#>  [5,] 0.8758659 0.8655931 0.5533034 0.6857547 0.6128834 0.5722686 0.6531523
#>  [6,] 0.8975256 0.8870015 0.6191397 0.7462232 0.6571028 0.6385429 0.7235807
#>  [7,] 0.8290630 0.8299941 0.4386136 0.5177113 0.5472317 0.5182554 0.6317049
#>  [8,] 0.8742926 0.8662296 0.4944783 0.6104338 0.6262463 0.5804937 0.6810618
#>  [9,] 0.8821145 0.8550903 0.5825493 0.6874224 0.6855950 0.6826682 0.7616373
#> [10,] 0.8033860 0.7852533 0.4593459 0.5667704 0.5123166 0.5156739 0.5725708
#> [11,] 0.8633398 0.8529043 0.4994147 0.5835557 0.6268522 0.5997618 0.6924380
#> [12,] 0.7643643 0.7771610 0.3541191 0.4532466 0.4193007 0.3449433 0.4899128
#> [13,] 0.8583384 0.8265623 0.5952816 0.6986433 0.6980075 0.6907717 0.7269349
#> [14,] 0.9068270 0.8954709 0.5228750 0.6419453 0.6295787 0.6506639 0.7326166
#> [15,] 0.7781460 0.7883809 0.4360792 0.4919839 0.5291930 0.4844666 0.6154917
#> [16,] 0.9236127 0.9077886 0.6265804 0.7331377 0.7132584 0.7288835 0.7771462
#> [17,] 0.9170209 0.9001770 0.6090327 0.7093606 0.6956401 0.7032113 0.7737947
#> [18,] 0.7540007 0.7816178 0.3738576 0.4667677 0.4216073 0.3181268 0.4827222
#> [19,] 0.8093328 0.8050559 0.5489620 0.6782657 0.5930448 0.5480027 0.6454191
#> [20,] 0.8553215 0.8320077 0.4929646 0.6642447 0.5494994 0.5092678 0.6104979
#> [21,] 0.9107998 0.9008213 0.5138216 0.6928343 0.5928019 0.5316803 0.6875062
#> [22,] 0.8417093 0.8518541 0.3843130 0.5113253 0.4938422 0.4115197 0.6092897
#> [23,] 0.8737854 0.8606116 0.5723528 0.7176820 0.6088319 0.5972251 0.6765308
#> [24,] 0.7797507 0.7765695 0.5110423 0.6378192 0.5201141 0.4855203 0.5777409
#> [25,] 0.7807517 0.7653703 0.4857101 0.5926844 0.4947990 0.4797745 0.5791697
#> [26,] 0.8073139 0.8032920 0.4869998 0.6281075 0.5263092 0.4560080 0.5599508
#> [27,] 0.7346579 0.7015366 0.3827708 0.4946263 0.4778553 0.4324139 0.5340064
#> [28,] 0.8553628 0.8385153 0.4840785 0.6463375 0.5429537 0.4938183 0.6052931
#> [29,] 0.8596325 0.8571466 0.4374155 0.5081373 0.5693650 0.5428881 0.6634412
#> [30,] 0.8800317 0.8531489 0.5662137 0.7011745 0.6495181 0.6364763 0.7080751
#> [31,] 0.7684741 0.7628206 0.4128811 0.5271934 0.5085954 0.4115978 0.5272027
#> [32,] 0.8510816 0.8264379 0.5250726 0.6628233 0.6028476 0.5891480 0.6507759
#> [33,] 0.8490502 0.8524383 0.4961157 0.5696806 0.5945638 0.5066589 0.6479560
#> [34,] 0.7862961 0.8025904 0.4017124 0.4657372 0.5315247 0.4768098 0.6165292
#> [35,] 0.8319822 0.8139499 0.4517381 0.6232980 0.4974074 0.4431108 0.5666098
#> [36,] 0.9077593 0.8990139 0.5311077 0.6276069 0.6571278 0.6487424 0.7328715
#> [37,] 0.9298834 0.9200607 0.6488446 0.7840094 0.6964259 0.6689525 0.7616472
#> [38,] 0.9068465 0.8951033 0.5667765 0.7180801 0.6434182 0.6223335 0.7267047
#> [39,] 0.8899629 0.8820363 0.5026941 0.5988725 0.6036225 0.5665601 0.6703361
#> [40,] 0.7757506 0.7602218 0.4829977 0.5491496 0.5699861 0.5308185 0.6215764
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7234533 0.7790387 0.6294231
#>  [2,] 0.5283541 0.6723589 0.5409315
#>  [3,] 0.6863534 0.7780733 0.6704330
#>  [4,] 0.6549764 0.6682008 0.4270862
#>  [5,] 0.6409811 0.7563737 0.6374028
#>  [6,] 0.7003551 0.8057173 0.6878758
#>  [7,] 0.6645166 0.6792394 0.4797958
#>  [8,] 0.6468995 0.7459169 0.5827957
#>  [9,] 0.7051470 0.8053640 0.6848154
#> [10,] 0.5496493 0.6708628 0.4999723
#> [11,] 0.6824375 0.7602172 0.5680151
#> [12,] 0.5513095 0.5507186 0.3618830
#> [13,] 0.6586581 0.7916057 0.7069595
#> [14,] 0.7430271 0.7972480 0.5906176
#> [15,] 0.6469432 0.6553954 0.4602386
#> [16,] 0.7194186 0.8396994 0.7325796
#> [17,] 0.7062330 0.8306810 0.7090274
#> [18,] 0.5610947 0.5160111 0.3672277
#> [19,] 0.6607555 0.7091354 0.6039741
#> [20,] 0.5503986 0.7102237 0.5910259
#> [21,] 0.6931977 0.7654777 0.6255910
#> [22,] 0.6071037 0.6751763 0.4400781
#> [23,] 0.6687463 0.7666997 0.6423446
#> [24,] 0.5976116 0.6754883 0.5557366
#> [25,] 0.5337165 0.6712532 0.5200316
#> [26,] 0.5695656 0.6606964 0.5696291
#> [27,] 0.4872419 0.6019758 0.4563137
#> [28,] 0.5491190 0.7024877 0.5825692
#> [29,] 0.6694413 0.7159561 0.4890891
#> [30,] 0.6560188 0.7860789 0.6652886
#> [31,] 0.5242189 0.6277139 0.4835222
#> [32,] 0.6117280 0.7382062 0.6023284
#> [33,] 0.5992017 0.7191223 0.5575462
#> [34,] 0.6404549 0.6549028 0.4460683
#> [35,] 0.5317175 0.6662549 0.5294803
#> [36,] 0.7221780 0.8062494 0.6177800
#> [37,] 0.7267029 0.8330150 0.7453275
#> [38,] 0.6921876 0.7952627 0.6727657
#> [39,] 0.6238684 0.7498556 0.6036534
#> [40,] 0.5899055 0.6707738 0.5266671
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4554701 0.5445299
#>  [2,] 0.4571677 0.5428323
#>  [3,] 0.4493808 0.5506192
#>  [4,] 0.4641108 0.5358892
#>  [5,] 0.4511299 0.5488701
#>  [6,] 0.4490037 0.5509963
#>  [7,] 0.4599061 0.5400939
#>  [8,] 0.4554388 0.5445612
#>  [9,] 0.4521921 0.5478079
#> [10,] 0.4600117 0.5399883
#> [11,] 0.4571407 0.5428593
#> [12,] 0.4629953 0.5370047
#> [13,] 0.4520529 0.5479471
#> [14,] 0.4543005 0.5456995
#> [15,] 0.4617245 0.5382755
#> [16,] 0.4493818 0.5506182
#> [17,] 0.4507812 0.5492188
#> [18,] 0.4614686 0.5385314
#> [19,] 0.4525142 0.5474858
#> [20,] 0.4533097 0.5466903
#> [21,] 0.4489041 0.5510959
#> [22,] 0.4597769 0.5402231
#> [23,] 0.4505356 0.5494644
#> [24,] 0.4548062 0.5451938
#> [25,] 0.4589354 0.5410646
#> [26,] 0.4532858 0.5467142
#> [27,] 0.4632565 0.5367435
#> [28,] 0.4536228 0.5463772
#> [29,] 0.4604585 0.5395415
#> [30,] 0.4516860 0.5483140
#> [31,] 0.4594481 0.5405519
#> [32,] 0.4543992 0.5456008
#> [33,] 0.4572516 0.5427484
#> [34,] 0.4624453 0.5375547
#> [35,] 0.4552415 0.5447585
#> [36,] 0.4541622 0.5458378
#> [37,] 0.4458670 0.5541330
#> [38,] 0.4494069 0.5505931
#> [39,] 0.4547804 0.5452196
#> [40,] 0.4604163 0.5395837
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]      [,2]         [,3]        [,4]        [,5]        [,6]
#>  [1,] 1.724607 1.8313958  0.338526535  0.40650349  0.62379635  0.54565929
#>  [2,] 1.652368 1.5346626 -0.107340892  0.35495920  0.08516887 -0.05364198
#>  [3,] 2.025187 1.9473852  0.378560869  1.01043483  0.57980052  0.49777212
#>  [4,] 1.389545 1.4596544 -0.355803101 -0.20468719  0.15610346  0.01934752
#>  [5,] 1.953850 1.8625428  0.214026990  0.78034605  0.45944864  0.29111298
#>  [6,] 2.170029 2.0604725  0.485898339  1.07856966  0.65040982  0.56904537
#>  [7,] 1.579001 1.5855853 -0.246790465  0.07087492  0.18949187  0.07305404
#>  [8,] 1.939458 1.8680253 -0.022087813  0.44913593  0.51614657  0.32480031
#>  [9,] 2.012608 1.7750966  0.333247239  0.78809614  0.77960512  0.76606125
#> [10,] 1.407593 1.2965473 -0.162976087  0.26868652  0.04927621  0.06271635
#> [11,] 1.843311 1.7575638 -0.002341029  0.33738698  0.51873610  0.40447262
#> [12,] 1.176758 1.2491982 -0.600982020 -0.18756158 -0.32564475 -0.64134169
#> [13,] 1.801557 1.5614571  0.385842935  0.84084590  0.83782752  0.80372975
#> [14,] 2.275494 2.1478843  0.091564009  0.58381719  0.53040976  0.62195891
#> [15,] 1.254895 1.3151935 -0.257090126 -0.03206704  0.11690494 -0.06215342
#> [16,] 2.492477 2.2869273  0.517574462  1.01060098  0.91126238  0.98896509
#> [17,] 2.402542 2.1991923  0.443248165  0.89228051  0.82662144  0.86263721
#> [18,] 1.120065 1.2751189 -0.515702747 -0.13312555 -0.31617856 -0.76239350
#> [19,] 1.445680 1.4181988  0.196477690  0.74581306  0.37656700  0.19260417
#> [20,] 1.776963 1.5999236 -0.028143467  0.68226779  0.19864842  0.03707563
#> [21,] 2.323440 2.2063841  0.055300588  0.81340352  0.37556066  0.12689102
#> [22,] 1.671002 1.7492176 -0.471281371  0.04530909 -0.02463251 -0.35768662
#> [23,] 1.934851 1.8203790  0.291457049  0.93299263  0.44240463  0.39391612
#> [24,] 1.264214 1.2457858  0.044176371  0.56591112  0.08049973 -0.05793484
#> [25,] 1.270052 1.1823511 -0.057175190  0.37507354 -0.02080457 -0.08094626
#> [26,] 1.432650 1.4069976 -0.052012507  0.52410650  0.10533398 -0.17642435
#> [27,] 1.018385 0.8546256 -0.477803980 -0.02149550 -0.08863678 -0.27200935
#> [28,] 1.777297 1.6472225 -0.063707446  0.60297882  0.17223930 -0.02472821
#> [29,] 1.812241 1.7917900 -0.251657844  0.03255195  0.27926064  0.17197518
#> [30,] 1.992731 1.7595146  0.266419698  0.85289704  0.61692145  0.56010276
#> [31,] 1.199715 1.1682059 -0.352067571  0.10888089  0.03438480 -0.35736410
#> [32,] 1.743110 1.5605892  0.100374753  0.67590142  0.41734449  0.36044447
#> [33,] 1.727171 1.7538546 -0.015537477  0.28054829  0.38286439  0.02663734
#> [34,] 1.302742 1.4025639 -0.398335217 -0.13726626  0.12626614 -0.09282738
#> [35,] 1.599741 1.4758830 -0.193650625  0.50357041 -0.01037047 -0.22854636
#> [36,] 2.286578 2.1863155  0.124591567  0.52196382  0.65052067  0.61351569
#> [37,] 2.584899 2.4431724  0.613964247  1.28918621  0.83033559  0.70345105
#> [38,] 2.275724 2.1439635  0.268711309  0.93495797  0.59023131  0.49946465
#> [39,] 2.090363 2.0118562  0.010776392  0.40076937  0.42058194  0.26783019
#> [40,] 1.241072 1.1538958 -0.068035497  0.19723513  0.28179454  0.12343065
#>              [,7]        [,8]       [,9]         [,10]
#>  [1,]  0.98645646  0.96165631 1.26007316  0.5297425789
#>  [2,]  0.33632813  0.11353813 0.71887348  0.1640931338
#>  [3,]  0.84521041  0.78312570 1.25447350  0.7101440842
#>  [4,]  0.47117882  0.64098659 0.70005898 -0.2937493807
#>  [5,]  0.63292456  0.57962490 1.13290030  0.5641091166
#>  [6,]  0.96229327  0.84898946 1.42241843  0.7902071000
#>  [7,]  0.53953781  0.68348746 0.75027869 -0.0808608090
#>  [8,]  0.75865574  0.60543839 1.07695283  0.3342607445
#>  [9,]  1.16167721  0.87192947 1.42016323  0.7759906040
#> [10,]  0.29234787  0.19925392 0.71209006 -0.0001108017
#> [11,]  0.81154195  0.76499644 1.15387049  0.2737574396
#> [12,] -0.04035439  0.20596286 0.20357438 -0.5672004949
#> [13,]  0.97912663  0.65731975 1.33463144  0.8806624795
#> [14,]  1.00793885  1.06176191 1.36918253  0.3665192323
#> [15,]  0.47045612  0.60562974 0.64284096 -0.1593820852
#> [16,]  1.24911259  0.94157928 1.65599334  1.0077500895
#> [17,]  1.22986376  0.87715835 1.59046166  0.8906650774
#> [18,] -0.06913874  0.24560621 0.06406645 -0.5441284477
#> [19,]  0.59896355  0.66666257 0.89118855  0.4220516265
#> [20,]  0.44940570  0.20228148 0.89647092  0.3682079204
#> [21,]  0.78848638  0.81511148 1.18294947  0.5133479285
#> [22,]  0.44432769  0.43515405 0.73169127 -0.2408452949
#> [23,]  0.73787430  0.70252043 1.18976849  0.5855550056
#> [24,]  0.31350635  0.39552316 0.73311409  0.2238768509
#> [25,]  0.31936601  0.13507086 0.71385876  0.0801694029
#> [26,]  0.24096232  0.28007886 0.66639929  0.2803382514
#> [27,]  0.13623607 -0.05104346 0.41370462 -0.1751920756
#> [28,]  0.42756946  0.19711174 0.85917251  0.3333291395
#> [29,]  0.67866746  0.70565920 0.92449018 -0.0436506996
#> [30,]  0.88605369  0.64560242 1.30144990  0.6869521951
#> [31,]  0.10891838  0.09695167 0.52242201 -0.0659349229
#> [32,]  0.62245166  0.45458174 1.03666600  0.4151760943
#> [33,]  0.61006651  0.40213978 0.94011201  0.2312093675
#> [34,]  0.47484208  0.57733915 0.64066049 -0.2165692027
#> [35,]  0.26803261  0.12704041 0.69129487  0.1180580897
#> [36,]  1.00924062  0.95529121 1.42582155  0.4801360974
#> [37,]  1.16173126  0.97795827 1.60714801  1.0738452009
#> [38,]  0.97796734  0.81036637 1.35694463  0.7207207221
#> [39,]  0.70970556  0.50600053 1.09784225  0.4207110782
#> [40,]  0.49624448  0.36357476 0.71168678  0.1067698719
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] -0.1306853 0.04790751
#>  [2,] -0.1296567 0.04209361
#>  [3,] -0.1471716 0.05600142
#>  [4,] -0.1166876 0.02711643
#>  [5,] -0.1424709 0.05363572
#>  [6,] -0.1472241 0.05747273
#>  [7,] -0.1254346 0.03528602
#>  [8,] -0.1296486 0.04907024
#>  [9,] -0.1362542 0.05556340
#> [10,] -0.1278117 0.03248395
#> [11,] -0.1272369 0.04462188
#> [12,] -0.1264635 0.02182660
#> [13,] -0.1360369 0.05634266
#> [14,] -0.1340942 0.04921523
#> [15,] -0.1255309 0.02787126
#> [16,] -0.1386282 0.06454060
#> [17,] -0.1362330 0.06128188
#> [18,] -0.1307401 0.02369151
#> [19,] -0.1470790 0.04343831
#> [20,] -0.1398153 0.04749180
#> [21,] -0.1485662 0.05653343
#> [22,] -0.1281044 0.03313665
#> [23,] -0.1467817 0.05172520
#> [24,] -0.1455319 0.03573799
#> [25,] -0.1341500 0.03047911
#> [26,] -0.1438213 0.04358208
#> [27,] -0.1235299 0.02370970
#> [28,] -0.1385024 0.04754112
#> [29,] -0.1203130 0.03818406
#> [30,] -0.1395189 0.05434211
#> [31,] -0.1304021 0.03216240
#> [32,] -0.1364090 0.04650249
#> [33,] -0.1268698 0.04454212
#> [34,] -0.1217599 0.02874213
#> [35,] -0.1392020 0.04031241
#> [36,] -0.1307092 0.05315850
#> [37,] -0.1504080 0.06697592
#> [38,] -0.1444468 0.05862047
#> [39,] -0.1285334 0.05284044
#> [40,] -0.1258726 0.03279417
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5445299 -0.5445299
#>  [2,]  0.5428323 -0.5428323
#>  [3,] -0.4493808  0.4493808
#>  [4,]  0.5358892 -0.5358892
#>  [5,]  0.5488701 -0.5488701
#>  [6,] -0.4490037  0.4490037
#>  [7,] -0.4599061  0.4599061
#>  [8,] -0.4554388  0.4554388
#>  [9,] -0.4521921  0.4521921
#> [10,]  0.5399883 -0.5399883
#> [11,] -0.4571407  0.4571407
#> [12,]  0.5370047 -0.5370047
#> [13,] -0.4520529  0.4520529
#> [14,] -0.4543005  0.4543005
#> [15,]  0.5382755 -0.5382755
#> [16,]  0.5506182 -0.5506182
#> [17,]  0.5492188 -0.5492188
#> [18,]  0.5385314 -0.5385314
#> [19,] -0.4525142  0.4525142
#> [20,]  0.5466903 -0.5466903
#> [21,] -0.4489041  0.4489041
#> [22,] -0.4597769  0.4597769
#> [23,] -0.4505356  0.4505356
#> [24,] -0.4548062  0.4548062
#> [25,]  0.5410646 -0.5410646
#> [26,] -0.4532858  0.4532858
#> [27,]  0.5367435 -0.5367435
#> [28,]  0.5463772 -0.5463772
#> [29,]  0.5395415 -0.5395415
#> [30,]  0.5483140 -0.5483140
#> [31,]  0.5405519 -0.5405519
#> [32,]  0.5456008 -0.5456008
#> [33,] -0.4572516  0.4572516
#> [34,] -0.4624453  0.4624453
#> [35,]  0.5447585 -0.5447585
#> [36,]  0.5458378 -0.5458378
#> [37,] -0.4458670  0.4458670
#> [38,] -0.4494069  0.4494069
#> [39,]  0.5452196 -0.5452196
#> [40,]  0.5395837 -0.5395837
#> 
#> $L
#> [1] 0.7013808 0.7132263 0.7221503 0.7164644 0.7215204 0.7026860
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4492754