Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12         V13          V14
#>  [1,] -0.3781692 -0.08727147 -0.184971060 -0.15043691 -0.04545703 -0.139430687
#>  [2,] -0.2782008 -0.12421283 -0.171568235 -0.12613587 -0.09066968 -0.124482368
#>  [3,] -0.2503844 -0.03222791 -0.120678326 -0.15645611  0.02832727 -0.066025385
#>  [4,] -0.3708791 -0.13438693 -0.186252280 -0.02854887 -0.14570058 -0.011580514
#>  [5,] -0.2608493 -0.18056555  0.005471805 -0.01368844 -0.12429063 -0.102365123
#>  [6,] -0.3972534 -0.04809643 -0.041188167 -0.10949966 -0.07920107  0.019183964
#>  [7,] -0.3641242 -0.10070574 -0.011543412 -0.11683790 -0.09686810 -0.038371009
#>  [8,] -0.2073349 -0.07391023 -0.139531854  0.01029331 -0.08275976  0.044137928
#>  [9,] -0.1782316 -0.14054163  0.040004033  0.03107511 -0.03029686  0.009407249
#> [10,] -0.2848920 -0.14524709 -0.118088362 -0.17347282 -0.01971029 -0.095590492
#>                V15          V16         V17           V18         V19
#>  [1,] -0.124953059 -0.114534140  0.07681719  0.0692761798  0.12018283
#>  [2,]  0.023020956  0.024453805  0.09756368  0.1488997703  0.03173335
#>  [3,] -0.189879358  0.083096563 -0.10928538 -0.0440195367  0.09338688
#>  [4,] -0.016050184  0.024616081  0.07134140 -0.0001972357  0.00421436
#>  [5,] -0.067826708  0.003983003 -0.08308475 -0.0181526465 -0.01402774
#>  [6,]  0.011695965  0.019414165  0.12770694  0.0884965670  0.12223746
#>  [7,]  0.041718228  0.087075328  0.08769386  0.0029725159  0.09870675
#>  [8,] -0.076032428 -0.042428325  0.10541511  0.0356041273 -0.01344791
#>  [9,]  0.102960916  0.052940764  0.08357068  0.1699981413  0.12760681
#> [10,] -0.004626352 -0.005815875 -0.04387086 -0.0250906131  0.04234370
#>               V2         V20         V21          V22         V23        V24
#>  [1,] -0.2316652  0.08313137  0.13340501  0.135880948  0.09332800 0.25621326
#>  [2,] -0.2013700  0.16681233  0.06115975  0.104240216  0.16791587 0.04196216
#>  [3,] -0.2961803 -0.06467975 -0.02265803  0.114118033  0.15575086 0.07974506
#>  [4,] -0.2539684  0.05286116 -0.01011739  0.032377207  0.04981943 0.16336462
#>  [5,] -0.3333291  0.02765456  0.06383820  0.121884275  0.08849458 0.11474138
#>  [6,] -0.3829978  0.08078472  0.14319212  0.205767933  0.08202209 0.21944022
#>  [7,] -0.2621380  0.11034572  0.09741483  0.038302305  0.17529051 0.12041960
#>  [8,] -0.1935370  0.03331892  0.10438900  0.122129246  0.03515864 0.15582710
#>  [9,] -0.3199415  0.10818870  0.08260054  0.123135677  0.10739535 0.08668474
#> [10,] -0.2188507  0.06635556  0.02381361 -0.003344031 -0.04898435 0.07841653
#>               V25        V26         V27        V28           V29         V3
#>  [1,]  0.13292049 0.14376221 0.129637314 0.18641211  0.0540654745 -0.3388674
#>  [2,]  0.05563624 0.17405011 0.203347458 0.13675867  0.0713110949 -0.2034439
#>  [3,] -0.03467372 0.01439312 0.007142195 0.10719783  0.0442497551 -0.2143624
#>  [4,]  0.15588420 0.05261883 0.141441795 0.09351743  0.0800097956 -0.2704963
#>  [5,]  0.15616018 0.29339386 0.222364676 0.12506407  0.0442329960 -0.2517869
#>  [6,]  0.19779770 0.30263246 0.237615003 0.20634046  0.1859002860 -0.3302404
#>  [7,]  0.28925401 0.20416011 0.128918180 0.21931896  0.2073127495 -0.3955684
#>  [8,]  0.12992285 0.09253214 0.090826569 0.09732678  0.0430799863 -0.2202503
#>  [9,]  0.09132567 0.09272348 0.042022182 0.09949769  0.0008175288 -0.2018446
#> [10,]  0.02098963 0.04499700 0.083078395 0.01051204 -0.0236294877 -0.2120959
#>               V30          V31          V32          V33          V34
#>  [1,]  0.11640564 -0.014825586 -0.000626876  0.003221724  0.055836644
#>  [2,]  0.03292069  0.018941774 -0.145499099 -0.146571715 -0.176063727
#>  [3,] -0.05644483 -0.048213327 -0.167959186 -0.060667176 -0.070949503
#>  [4,]  0.08359463  0.080823713  0.008163588 -0.125156066 -0.146586999
#>  [5,]  0.06404022 -0.022109650 -0.010035586  0.029552828  0.005540993
#>  [6,]  0.09576774  0.070781102  0.002578813  0.033231962  0.014209602
#>  [7,]  0.04186894  0.056680330  0.078056977  0.052680612 -0.004602407
#>  [8,]  0.10071887  0.086756447  0.038961203  0.001813063 -0.084408263
#>  [9,]  0.06218086 -0.020887531 -0.029103206  0.024070159 -0.096270071
#> [10,]  0.05996058 -0.007461037 -0.010124829 -0.047652866  0.025488158
#>                V35         V36          V37          V38         V39         V4
#>  [1,] -0.082770498 -0.05846779  0.029160666 -0.004356147 -0.02212624 -0.3523792
#>  [2,] -0.038023757 -0.13442515 -0.006338108 -0.170421141 -0.20983993 -0.3022491
#>  [3,] -0.099095924 -0.09555352 -0.163372553 -0.208933121 -0.10418124 -0.2903268
#>  [4,] -0.111232589 -0.08443913 -0.126956701 -0.025353442 -0.10121705 -0.1871167
#>  [5,] -0.025277464 -0.05157541 -0.107417890 -0.118931279 -0.07321993 -0.2686262
#>  [6,] -0.007183253 -0.09800873 -0.037341081 -0.126080420 -0.05294380 -0.3623737
#>  [7,]  0.069250490  0.10431991  0.006264147  0.088573843  0.04211149 -0.2612853
#>  [8,] -0.009128829  0.08235659  0.185161647  0.089869862  0.10403284 -0.2630572
#>  [9,] -0.074810101 -0.07215259 -0.133182727 -0.162419263 -0.08854794 -0.2214661
#> [10,] -0.091571787  0.05892249 -0.062997180  0.027962396  0.04114277 -0.2895929
#>               V40         V41         V42         V43         V44         V45
#>  [1,]  0.02820049 -0.06258182 -0.01179572 -0.02940568 -0.14959619 -0.16858624
#>  [2,] -0.18759488 -0.11353271 -0.01687630 -0.21370449 -0.07590180 -0.14391911
#>  [3,] -0.18013495 -0.05656724 -0.15406046 -0.17667513 -0.11428109 -0.21312379
#>  [4,] -0.17672667 -0.21776751 -0.04687538 -0.12348292 -0.21559284 -0.15582338
#>  [5,] -0.01249771 -0.05395615 -0.14254473 -0.11822666 -0.06940570 -0.13421460
#>  [6,] -0.16773188 -0.06587865 -0.15116995 -0.10127233 -0.17235619 -0.10504461
#>  [7,] -0.00842413  0.01387379  0.04244269 -0.10251673  0.01501553 -0.02649679
#>  [8,]  0.06122694  0.07290555  0.03315285 -0.09279857 -0.14350634 -0.10609232
#>  [9,]  0.07242129  0.04643822 -0.10363388 -0.05599378 -0.10187859 -0.03795782
#> [10,] -0.05423346 -0.06934609 -0.14977420 -0.04546017 -0.13346625 -0.19610080
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.16076828 -0.16886432 -0.2527436 -0.2394962 -0.1665134 -0.2366686
#>  [2,] -0.26445699 -0.19750669 -0.1490603 -0.3571640 -0.2018160 -0.4080365
#>  [3,] -0.10196475 -0.13722611 -0.2059532 -0.2726844 -0.1693287 -0.2836776
#>  [4,] -0.14594328 -0.25493438 -0.1713337 -0.2472671 -0.2555171 -0.2724008
#>  [5,] -0.14624062 -0.14460436 -0.1805028 -0.2237767 -0.2922659 -0.4053392
#>  [6,] -0.14609264 -0.26722555 -0.2586791 -0.3119129 -0.3507549 -0.3283765
#>  [7,] -0.03903698 -0.15076243 -0.2170474 -0.2643753 -0.2197512 -0.2956016
#>  [8,] -0.02559459 -0.18617984 -0.2362303 -0.2155937 -0.1331153 -0.3617915
#>  [9,] -0.21697529 -0.07193833 -0.2454183 -0.2715343 -0.1740305 -0.3752211
#> [10,] -0.16283984 -0.14601187 -0.2174541 -0.3032501 -0.2560555 -0.1838085
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2206807 -0.2754496 -0.3121309 -0.3496798 -0.3543884 -0.2904440
#>  [2,] -0.2695405 -0.2694585 -0.3115877 -0.3573038 -0.3381282 -0.3541987
#>  [3,] -0.3688248 -0.3442697 -0.2480049 -0.3735134 -0.3367414 -0.3895308
#>  [4,] -0.3946069 -0.3014754 -0.2898712 -0.3902870 -0.2703318 -0.4287614
#>  [5,] -0.3031692 -0.3684233 -0.4051180 -0.3911636 -0.4229717 -0.2851375
#>  [6,] -0.4142031 -0.3623055 -0.4496471 -0.3632115 -0.3443875 -0.3645421
#>  [7,] -0.2567815 -0.4153943 -0.4336872 -0.3567475 -0.3059486 -0.3324215
#>  [8,] -0.3550310 -0.3888896 -0.2599214 -0.2596176 -0.3760334 -0.2791856
#>  [9,] -0.3578115 -0.3208373 -0.3036806 -0.3175826 -0.2811569 -0.3764472
#> [10,] -0.2890782 -0.3066512 -0.2141665 -0.2888321 -0.3670984 -0.2096273
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.2999037 -0.4214420 -0.2484703 -0.27943797 -0.3909845 -0.12206696
#>  [2,] -0.2573916 -0.2725180 -0.2393837 -0.26775500 -0.3566499 -0.22091338
#>  [3,] -0.3195057 -0.2911869 -0.2252195 -0.21948578 -0.2490332 -0.18998315
#>  [4,] -0.3395736 -0.2914039 -0.3475114 -0.15841998 -0.3811660 -0.14692571
#>  [5,] -0.4429158 -0.3761376 -0.4269717 -0.09637381 -0.2729121 -0.12940522
#>  [6,] -0.3611378 -0.4046083 -0.3110625 -0.16312757 -0.5054091 -0.20565708
#>  [7,] -0.3561615 -0.3157351 -0.4287479 -0.25279925 -0.3483618 -0.07256361
#>  [8,] -0.2879724 -0.3990687 -0.2804491 -0.18212349 -0.2393240 -0.25981935
#>  [9,] -0.3516522 -0.2669344 -0.3473960 -0.11597446 -0.3375781 -0.20809145
#> [10,] -0.2292294 -0.2448326 -0.2985412 -0.18694528 -0.3387912 -0.12538680
#>                V8          V9
#>  [1,] -0.08369291 -0.18289588
#>  [2,] -0.19625736 -0.19393663
#>  [3,] -0.19307556 -0.07110800
#>  [4,] -0.21120262 -0.23190582
#>  [5,] -0.25647110 -0.08624635
#>  [6,] -0.26185825 -0.16395745
#>  [7,] -0.14119355 -0.03284783
#>  [8,] -0.14382183 -0.18957518
#>  [9,] -0.20113338 -0.06015698
#> [10,] -0.16871788 -0.18485828
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,]  0.05078572 -0.01357055 -0.07616625 0.05831046  0.10031005  0.09190865
#> [2,] -0.01282516  0.03357034 -0.02810190 0.06929643 -0.01076238 -0.03574794
#>            [,7]        [,8]        [,9]      [,10]
#> [1,]  0.1801379  0.12595031  0.09788286 0.07133870
#> [2,] -0.1438523 -0.05744723 -0.16292859 0.08387957
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  8.642772e-05  0.0007222721  0.0008142429  0.0008090899  0.0007582661
#>  [2,]  2.204639e-04  0.0015794660  0.0019771998  0.0019293807  0.0015933802
#>  [3,]  1.350291e-04  0.0010211263  0.0013086692  0.0013287298  0.0010694727
#>  [4,]  1.506434e-04  0.0011177748  0.0013777457  0.0013610472  0.0011563517
#>  [5,]  3.298392e-05  0.0004120817  0.0003979216  0.0004275067  0.0004843122
#>  [6,]  4.328581e-05  0.0004539377  0.0004422745  0.0004424159  0.0004806725
#>  [7,] -1.651022e-04 -0.0006557329 -0.0009915166 -0.0008011230 -0.0004504660
#>  [8,] -8.909108e-05 -0.0003247211 -0.0005470283 -0.0004426230 -0.0001851460
#>  [9,] -1.691513e-04 -0.0008291294 -0.0012159444 -0.0010443472 -0.0006161385
#> [10,]  1.176168e-04  0.0008993183  0.0010932684  0.0010820343  0.0009329949
#>                 V14          V15          V16          V17          V18
#>  [1,]  6.007202e-04 5.311199e-04 0.0005141686 0.0005081898 0.0006128516
#>  [2,]  1.110175e-03 7.657535e-04 0.0005542374 0.0005252720 0.0009956450
#>  [3,]  7.049220e-04 4.440390e-04 0.0002899344 0.0002994833 0.0006986881
#>  [4,]  8.439146e-04 6.229655e-04 0.0005117887 0.0005058629 0.0008206907
#>  [5,]  4.263265e-04 4.567032e-04 0.0005160949 0.0005366009 0.0005221770
#>  [6,]  4.040286e-04 4.281389e-04 0.0004607163 0.0004449072 0.0004405505
#>  [7,] -2.602464e-04 1.987625e-05 0.0003069534 0.0004466361 0.0001321044
#>  [8,] -2.520169e-05 1.904462e-04 0.0004154225 0.0005064607 0.0002576185
#>  [9,] -3.043268e-04 7.418998e-05 0.0004375585 0.0005684150 0.0001460113
#> [10,]  7.088729e-04 5.475868e-04 0.0004832216 0.0004800484 0.0007311945
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0009860723  1.087321e-04  0.0014202112  0.0015801319  0.0015218787
#>  [2,]  0.0024244224  2.804080e-04  0.0033940201  0.0036099570  0.0034902432
#>  [3,]  0.0016732097  1.779851e-04  0.0022345151  0.0023483551  0.0023336869
#>  [4,]  0.0017473764  1.911255e-04  0.0024176775  0.0025845045  0.0025221992
#>  [5,]  0.0005085373  4.388334e-05  0.0007068590  0.0008475739  0.0008287497
#>  [6,]  0.0004811685  5.543070e-05  0.0007252035  0.0008581592  0.0007958101
#>  [7,] -0.0010694232 -1.952121e-04 -0.0016725422 -0.0015914547 -0.0014618841
#>  [8,] -0.0005402794 -1.068900e-04 -0.0008339480 -0.0007433558 -0.0006827987
#>  [9,] -0.0013068303 -2.050431e-04 -0.0019818145 -0.0019079418 -0.0017883333
#> [10,]  0.0014297014  1.493998e-04  0.0019593033  0.0020996511  0.0020537827
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0013967990  0.0013448992  0.0012316515  1.342640e-03  0.0014084477
#>  [2,]  0.0031235298  0.0029441220  0.0026087071  2.516647e-03  0.0029764347
#>  [3,]  0.0021157123  0.0019709837  0.0016989088  1.572622e-03  0.0018945641
#>  [4,]  0.0023001006  0.0021908230  0.0019642481  1.933479e-03  0.0022088263
#>  [5,]  0.0007845106  0.0007639555  0.0007274417  9.103975e-04  0.0008355861
#>  [6,]  0.0007323587  0.0007212590  0.0006830311  8.495934e-04  0.0008021733
#>  [7,] -0.0012276634 -0.0011262019 -0.0008577491 -4.543674e-04 -0.0009579803
#>  [8,] -0.0005485863 -0.0004780315 -0.0003132225 -9.148074e-06 -0.0003813988
#>  [9,] -0.0014509877 -0.0012890761 -0.0010486059 -6.636817e-04 -0.0012865090
#> [10,]  0.0018948707  0.0018317990  0.0016670909  1.651036e-03  0.0018459284
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0014434261  0.0012771816  1.046949e-04  0.0010906267  7.035720e-04
#>  [2,]  0.0031902805  0.0027250726  2.550965e-04  0.0020867639  1.141099e-03
#>  [3,]  0.0020543952  0.0017815956  1.631987e-04  0.0013543769  7.491627e-04
#>  [4,]  0.0023193440  0.0020003404  1.789045e-04  0.0015722399  9.187122e-04
#>  [5,]  0.0008269988  0.0007758427  4.634003e-05  0.0007351397  5.464852e-04
#>  [6,]  0.0008324303  0.0007841754  5.342111e-05  0.0007541990  5.457099e-04
#>  [7,] -0.0011075142 -0.0008234944 -1.778776e-04 -0.0004793131 -3.486188e-05
#>  [8,] -0.0005286102 -0.0003678341 -8.732341e-05 -0.0001344275  9.835697e-05
#>  [9,] -0.0015207410 -0.0011274425 -1.685105e-04 -0.0006177088 -8.151494e-06
#> [10,]  0.0019140817  0.0016564718  1.434865e-04  0.0013292326  8.238216e-04
#>                V32          V33          V34           V35           V36
#>  [1,] 7.021462e-04 0.0007314009 0.0005869082  3.618693e-04  3.776099e-04
#>  [2,] 1.104739e-03 0.0009771538 0.0005609969 -1.051965e-04 -2.794450e-04
#>  [3,] 7.178269e-04 0.0006044203 0.0002660005 -1.855518e-04 -2.772796e-04
#>  [4,] 8.825471e-04 0.0008109602 0.0005274808  8.889435e-05 -5.088844e-06
#>  [5,] 5.687165e-04 0.0006547868 0.0005924408  5.190333e-04  6.217328e-04
#>  [6,] 5.784525e-04 0.0006643190 0.0006062278  5.238321e-04  6.132160e-04
#>  [7,] 6.703883e-05 0.0003078231 0.0004998810  9.004409e-04  1.280078e-03
#>  [8,] 1.460445e-04 0.0003333067 0.0004816906  7.305451e-04  9.777511e-04
#>  [9,] 8.500487e-05 0.0003567948 0.0006590645  1.168514e-03  1.563166e-03
#> [10,] 7.811976e-04 0.0007307678 0.0004992116  1.705848e-04  1.049625e-04
#>                 V37          V38          V39            V4          V40
#>  [1,]  4.450893e-04 0.0006372968 0.0007492676  1.506373e-04 0.0005138316
#>  [2,] -1.625109e-04 0.0005521002 0.0011721142  4.117904e-04 0.0005792587
#>  [3,] -1.528048e-04 0.0003407376 0.0007772361  2.619231e-04 0.0004172951
#>  [4,]  6.428569e-05 0.0005174164 0.0009034294  2.810560e-04 0.0005153552
#>  [5,]  6.980380e-04 0.0007115010 0.0006523131  5.223157e-05 0.0005380033
#>  [6,]  6.745439e-04 0.0007120539 0.0006769648  6.182204e-05 0.0005289540
#>  [7,]  1.400536e-03 0.0008691272 0.0003529883 -3.099234e-04 0.0006315094
#>  [8,]  1.041670e-03 0.0006724142 0.0002973713 -1.674343e-04 0.0004770853
#>  [9,]  1.607905e-03 0.0009143552 0.0002203746 -3.228262e-04 0.0006518727
#> [10,]  1.539097e-04 0.0004810606 0.0007577772  2.200070e-04 0.0004767262
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0005164617 0.0006708149  6.640518e-04  0.0006649351  0.0008154431
#>  [2,] 0.0007117946 0.0010687752  1.226869e-03  0.0014031956  0.0017531428
#>  [3,] 0.0004854846 0.0007004817  7.814649e-04  0.0008862882  0.0010655579
#>  [4,] 0.0005879939 0.0008332581  9.185251e-04  0.0010159968  0.0012333844
#>  [5,] 0.0004910252 0.0005595215  4.755364e-04  0.0003977061  0.0004748731
#>  [6,] 0.0004865101 0.0005469643  4.700898e-04  0.0004100826  0.0005107230
#>  [7,] 0.0003889764 0.0001467910 -1.541348e-04 -0.0004662730 -0.0006712632
#>  [8,] 0.0003085374 0.0001914450  8.274259e-06 -0.0002065281 -0.0003100851
#>  [9,] 0.0004122125 0.0001294512 -2.355866e-04 -0.0006276444 -0.0008979346
#> [10,] 0.0005088908 0.0006832612  7.367232e-04  0.0008051179  0.0009591025
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0005678602  0.0004089443  0.0003375352  1.934590e-04  1.719587e-04
#>  [2,]  0.0012384319  0.0010145012  0.0007693253  4.441414e-04  5.034590e-04
#>  [3,]  0.0007538938  0.0006507405  0.0004902253  2.838248e-04  3.465185e-04
#>  [4,]  0.0008656520  0.0006917844  0.0005298513  3.052116e-04  3.485350e-04
#>  [5,]  0.0003293606  0.0002013066  0.0001911787  1.112682e-04  5.257929e-05
#>  [6,]  0.0003622448  0.0002324922  0.0002127081  1.238101e-04  6.177060e-05
#>  [7,] -0.0004812386 -0.0004819287 -0.0003084256 -1.640758e-04 -3.568679e-04
#>  [8,] -0.0002342236 -0.0002771697 -0.0001629435 -8.746537e-05 -2.099142e-04
#>  [9,] -0.0006479036 -0.0006356765 -0.0004130242 -2.287116e-04 -3.876523e-04
#> [10,]  0.0006711585  0.0005333170  0.0004119856  2.358934e-04  2.769420e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  5.052388e-05  5.219586e-05  3.946214e-05  2.296852e-05  2.847864e-05
#>  [2,]  9.524504e-05  1.171447e-04  1.005390e-04  5.909888e-05  7.490061e-05
#>  [3,]  5.817850e-05  7.268729e-05  6.307510e-05  4.000290e-05  4.814622e-05
#>  [4,]  6.862508e-05  8.146072e-05  6.955262e-05  4.177480e-05  5.167237e-05
#>  [5,]  3.729004e-05  2.946721e-05  1.647339e-05  9.963641e-06  1.111954e-05
#>  [6,]  3.978747e-05  3.242747e-05  1.934277e-05  1.180165e-05  1.411713e-05
#>  [7,] -8.568135e-06 -5.013470e-05 -6.317498e-05 -3.251426e-05 -4.797722e-05
#>  [8,]  1.698692e-06 -2.805428e-05 -3.598273e-05 -1.936808e-05 -2.792731e-05
#>  [9,] -2.074996e-05 -6.314957e-05 -7.011315e-05 -3.733968e-05 -5.414411e-05
#> [10,]  5.492915e-05  6.321509e-05  5.382456e-05  3.367941e-05  4.135338e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  1.882712e-05  1.755634e-05  1.494745e-05  1.740921e-05  1.884869e-05
#>  [2,]  2.948891e-05  3.073604e-05  2.657006e-05  3.381563e-05  4.623900e-05
#>  [3,]  1.449934e-05  1.929638e-05  1.459603e-05  2.088989e-05  2.957438e-05
#>  [4,]  2.243093e-05  2.325135e-05  2.025879e-05  2.463868e-05  3.243726e-05
#>  [5,]  1.429019e-05  1.312329e-05  9.696216e-06  1.153880e-05  9.148726e-06
#>  [6,]  1.499948e-05  1.299828e-05  1.034955e-05  1.247005e-05  1.083429e-05
#>  [7,] -5.465948e-06 -3.219780e-06 -1.115419e-05 -1.097253e-05 -2.663375e-05
#>  [8,]  2.121203e-06  1.885996e-06 -2.806020e-06 -3.425032e-06 -1.449356e-05
#>  [9,] -2.858051e-07 -3.793088e-06 -9.605186e-06 -9.849657e-06 -2.926997e-05
#> [10,]  1.873687e-05  1.974893e-05  1.718392e-05  1.994130e-05  2.605160e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  2.323341e-04  1.225767e-05  2.643258e-04  3.530726e-04  0.0006173411
#>  [2,]  4.752280e-04  2.423163e-05  5.114196e-04  6.305219e-04  0.0013424264
#>  [3,]  3.218394e-04  1.544679e-05  3.345435e-04  4.027003e-04  0.0008509283
#>  [4,]  3.505576e-04  1.786683e-05  3.756838e-04  4.669178e-04  0.0009506205
#>  [5,]  1.449224e-04  8.108936e-06  1.793274e-04  2.527296e-04  0.0003418790
#>  [6,]  1.483743e-04  8.745346e-06  1.876936e-04  2.707764e-04  0.0003838355
#>  [7,] -1.636292e-04 -8.523658e-06 -1.179402e-04 -1.050381e-04 -0.0006233571
#>  [8,] -6.623077e-05 -3.083697e-06 -3.136312e-05 -1.062937e-05 -0.0003042312
#>  [9,] -1.782940e-04 -8.161214e-06 -1.470934e-04 -1.471595e-04 -0.0007503760
#> [10,]  2.900510e-04  1.508539e-05  3.133457e-04  3.913148e-04  0.0007663000
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.03477566 -0.03122018 -0.01765996 -0.02695642 -0.03786643 -0.0460564
#> [2,]  0.03477566  0.03122018  0.01765996  0.02695642  0.03786643  0.0460564
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.04352652 -0.03215138 -0.03797422 -0.01805621
#> [2,]  0.04352652  0.03215138  0.03797422  0.01805621
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.1678758 0.3559444 0.1106561 0.4504055 0.5779423 0.6669547 0.6064896
#>  [8] 0.4573050 0.4886062 0.4563516
#> 
#> $B[[2]]
#> [1]  0.02005169 -0.17302913
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0017551973  0.0029558734  0.0018507715  0.0023186494  0.0013433291
#>  [6]  0.0012906128 -0.0001026727  0.0003097664 -0.0001665388  0.0020060465
#> 
#> $vB[[2]]
#> [1] -0.05121599  0.05121599
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#>  [2,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#>  [3,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#>  [4,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [5,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#>  [6,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#>  [7,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#>  [8,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#>  [9,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [10,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [11,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [12,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [13,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [14,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [15,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [16,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [17,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [18,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [19,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [20,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [21,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [22,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [23,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [24,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [25,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [26,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [27,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [28,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [29,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [30,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [31,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [32,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [33,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [34,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [35,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [36,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [37,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [38,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [39,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [40,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#>  [2,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#>  [3,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#>  [4,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [5,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#>  [6,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#>  [7,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#>  [8,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#>  [9,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [10,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [11,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [12,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [13,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [14,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [15,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [16,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [17,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [18,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [19,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [20,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [21,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [22,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [23,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [24,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [25,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [26,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [27,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [28,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [29,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [30,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [31,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [32,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [33,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [34,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [35,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [36,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [37,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [38,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [39,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [40,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#>  [2,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#>  [3,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#>  [4,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [5,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#>  [6,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#>  [7,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#>  [8,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#>  [9,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [10,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [11,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [12,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [13,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [14,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [15,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [16,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [17,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [18,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [19,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [20,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [21,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [22,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [23,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [24,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [25,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [26,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [27,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [28,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [29,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [30,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [31,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [32,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [33,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [34,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [35,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [36,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [37,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [38,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [39,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [40,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#>  [2,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#>  [3,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#>  [4,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [5,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#>  [6,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#>  [7,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#>  [8,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#>  [9,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [10,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [11,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [12,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [13,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [14,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [15,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [16,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [17,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [18,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [19,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [20,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [21,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [22,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [23,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [24,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [25,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [26,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [27,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [28,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [29,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [30,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [31,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [32,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [33,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [34,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [35,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [36,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [37,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [38,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [39,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [40,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#>  [2,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#>  [3,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#>  [4,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [5,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#>  [6,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#>  [7,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#>  [8,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#>  [9,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [10,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [11,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [12,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [13,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [14,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [15,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [16,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [17,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [18,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [19,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [20,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [21,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [22,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [23,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [24,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [25,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [26,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [27,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [28,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [29,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [30,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [31,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [32,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [33,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [34,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [35,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [36,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [37,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [38,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [39,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [40,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#>  [2,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#>  [3,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#>  [4,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [5,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#>  [6,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#>  [7,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#>  [8,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#>  [9,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [10,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [11,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [12,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [13,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [14,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [15,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [16,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [17,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [18,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [19,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [20,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [21,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [22,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [23,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [24,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [25,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [26,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [27,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [28,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [29,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [30,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [31,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [32,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [33,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [34,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [35,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [36,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [37,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [38,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [39,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [40,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7779118 0.7788248 0.4924828 0.7226680 0.7958315 0.9183091 0.8977714
#>  [2,] 0.6812482 0.6541895 0.4007993 0.6136458 0.7208976 0.8583113 0.8522332
#>  [3,] 0.6742116 0.6078525 0.4226873 0.6194729 0.7435629 0.8469042 0.8448231
#>  [4,] 0.5702387 0.5329344 0.3151632 0.4726982 0.5514201 0.7321700 0.7728303
#>  [5,] 0.7030271 0.6240733 0.3642386 0.5667250 0.6802297 0.8352411 0.8593484
#>  [6,] 0.6106998 0.6248210 0.3593834 0.5581520 0.6451314 0.8287422 0.8247697
#>  [7,] 0.6787041 0.5775647 0.3655426 0.5928510 0.7221538 0.8449954 0.8520720
#>  [8,] 0.6731563 0.5849632 0.3390726 0.5412377 0.6528540 0.8024012 0.8418830
#>  [9,] 0.6381692 0.5881892 0.3332464 0.5489731 0.6613772 0.8161533 0.8501978
#> [10,] 0.6700264 0.6278855 0.4507912 0.6125754 0.7126555 0.8147141 0.8233813
#> [11,] 0.6670130 0.5513607 0.3298689 0.5213497 0.6905414 0.7947667 0.8585938
#> [12,] 0.6177711 0.5404301 0.3533103 0.5513407 0.6787367 0.7930771 0.8188461
#> [13,] 0.7257514 0.6906014 0.4084467 0.6493935 0.7466555 0.8823364 0.8761413
#> [14,] 0.7346056 0.6656335 0.3944053 0.6121745 0.7309253 0.8595441 0.8725033
#> [15,] 0.5415139 0.3622935 0.2474975 0.4051936 0.5561500 0.6342998 0.7729089
#> [16,] 0.4346929 0.3325632 0.1999511 0.3572807 0.4878224 0.6036449 0.7465035
#> [17,] 0.7486934 0.7028798 0.4663883 0.6712177 0.7921615 0.8951787 0.8792781
#> [18,] 0.7451427 0.7060979 0.4298220 0.6667542 0.7719736 0.8975341 0.8854168
#> [19,] 0.7047993 0.5641952 0.3267777 0.5297907 0.7247335 0.8358226 0.8851279
#> [20,] 0.6584686 0.6442462 0.4405752 0.6459579 0.7526753 0.8298111 0.8272504
#> [21,] 0.6898345 0.5746889 0.3288950 0.5550698 0.7234994 0.8294842 0.8586071
#> [22,] 0.7248635 0.5916904 0.3682392 0.5968771 0.7548725 0.8558492 0.8831566
#> [23,] 0.6623712 0.5551392 0.3210558 0.5255850 0.6936183 0.8229859 0.8468096
#> [24,] 0.6416917 0.5903845 0.4195975 0.5564728 0.6995306 0.8011642 0.8173892
#> [25,] 0.6646648 0.5992855 0.3904892 0.6140058 0.7319936 0.8361390 0.8455708
#> [26,] 0.6381900 0.6410909 0.4208221 0.6055927 0.6897484 0.8259641 0.8171266
#> [27,] 0.6979798 0.6685580 0.3758655 0.6027532 0.7155561 0.8738199 0.8719133
#> [28,] 0.6219335 0.5078055 0.2987858 0.4843741 0.5877590 0.7422363 0.8181817
#> [29,] 0.5237695 0.3376621 0.1842355 0.3638104 0.4962996 0.6370928 0.8004600
#> [30,] 0.6740105 0.6388804 0.4574267 0.6362600 0.7557575 0.8244080 0.8286701
#> [31,] 0.7447944 0.7415233 0.4611224 0.6834036 0.7695269 0.8947232 0.8727598
#> [32,] 0.7192743 0.6572076 0.4452076 0.6740270 0.7858399 0.8796891 0.8705282
#> [33,] 0.7178275 0.5453525 0.3449589 0.5381139 0.7250025 0.8469452 0.8711183
#> [34,] 0.6806253 0.6646301 0.3728410 0.6106146 0.6970462 0.8731004 0.8602162
#> [35,] 0.4439528 0.2253806 0.1576062 0.2777923 0.4185288 0.4889682 0.7372436
#> [36,] 0.4602814 0.4173495 0.2642408 0.4273713 0.5006651 0.6100460 0.7081088
#> [37,] 0.6962196 0.6363366 0.4295304 0.6429925 0.7642857 0.8612033 0.8528506
#> [38,] 0.5449752 0.5536830 0.3143167 0.5298394 0.5853091 0.7710752 0.7995139
#> [39,] 0.6698500 0.6081362 0.3924442 0.6099686 0.7247038 0.8365291 0.8432238
#> [40,] 0.7265557 0.6211339 0.4211843 0.6326343 0.7766599 0.8675077 0.8796570
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7800557 0.7970195 0.6147095
#>  [2,] 0.7239480 0.7434467 0.5155929
#>  [3,] 0.7145172 0.6824330 0.5215110
#>  [4,] 0.6897030 0.7089583 0.4869629
#>  [5,] 0.7602297 0.7369313 0.5445218
#>  [6,] 0.6852267 0.7586743 0.4811540
#>  [7,] 0.7428174 0.6951974 0.5267060
#>  [8,] 0.7465067 0.7053441 0.5352112
#>  [9,] 0.7489469 0.6962801 0.4888143
#> [10,] 0.7037820 0.6794128 0.5621309
#> [11,] 0.7631265 0.6118991 0.5404558
#> [12,] 0.6871215 0.6436478 0.4964902
#> [13,] 0.7584905 0.7594004 0.5453151
#> [14,] 0.7646663 0.7369448 0.5489450
#> [15,] 0.7181253 0.5127469 0.5205730
#> [16,] 0.5797905 0.4938845 0.3601478
#> [17,] 0.7498434 0.7493520 0.5753536
#> [18,] 0.7586847 0.7657127 0.5562031
#> [19,] 0.7826441 0.6458869 0.5387485
#> [20,] 0.7110172 0.6766769 0.5822361
#> [21,] 0.7543606 0.6730990 0.5151343
#> [22,] 0.7868751 0.6645837 0.5818678
#> [23,] 0.7263027 0.6653098 0.5242498
#> [24,] 0.6973966 0.6821477 0.5179195
#> [25,] 0.7211137 0.6767432 0.5393772
#> [26,] 0.7019165 0.6980814 0.5266647
#> [27,] 0.7415475 0.7519705 0.5048276
#> [28,] 0.7463744 0.6500009 0.5018238
#> [29,] 0.7232926 0.5518572 0.4575203
#> [30,] 0.7185492 0.6814833 0.5852420
#> [31,] 0.7534493 0.7877700 0.5904558
#> [32,] 0.7502005 0.7126728 0.5623154
#> [33,] 0.7588024 0.7219917 0.5312691
#> [34,] 0.7250184 0.7591972 0.5159169
#> [35,] 0.6634208 0.4333441 0.3895499
#> [36,] 0.6095577 0.5623678 0.4080908
#> [37,] 0.7219884 0.6959634 0.5363325
#> [38,] 0.6913900 0.7074556 0.4964948
#> [39,] 0.7164866 0.6844140 0.5332957
#> [40,] 0.7632860 0.6794817 0.5746353
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5978536 0.4021464
#>  [2,] 0.5982427 0.4017573
#>  [3,] 0.5955637 0.4044363
#>  [4,] 0.5962913 0.4037087
#>  [5,] 0.6007402 0.3992598
#>  [6,] 0.5983234 0.4016766
#>  [7,] 0.5994489 0.4005511
#>  [8,] 0.5995217 0.4004783
#>  [9,] 0.6003547 0.3996453
#> [10,] 0.5918982 0.4081018
#> [11,] 0.5983798 0.4016202
#> [12,] 0.5951676 0.4048324
#> [13,] 0.5996440 0.4003560
#> [14,] 0.5997519 0.4002481
#> [15,] 0.5941415 0.4058585
#> [16,] 0.5914982 0.4085018
#> [17,] 0.5974164 0.4025836
#> [18,] 0.5995409 0.4004591
#> [19,] 0.6021528 0.3978472
#> [20,] 0.5917065 0.4082935
#> [21,] 0.6005659 0.3994341
#> [22,] 0.6002897 0.3997103
#> [23,] 0.5995016 0.4004984
#> [24,] 0.5943237 0.4056763
#> [25,] 0.5961140 0.4038860
#> [26,] 0.5930468 0.4069532
#> [27,] 0.6007237 0.3992763
#> [28,] 0.5988329 0.4011671
#> [29,] 0.6004412 0.3995588
#> [30,] 0.5920960 0.4079040
#> [31,] 0.5973859 0.4026141
#> [32,] 0.5968902 0.4031098
#> [33,] 0.6042136 0.3957864
#> [34,] 0.5996804 0.4003196
#> [35,] 0.5936707 0.4063293
#> [36,] 0.5881364 0.4118636
#> [37,] 0.5956753 0.4043247
#> [38,] 0.5969172 0.4030828
#> [39,] 0.5959843 0.4040157
#> [40,] 0.5980513 0.4019487
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]        [,2]        [,3]        [,4]         [,5]        [,6]
#>  [1,]  1.25353829  1.25883093 -0.03007094  0.95773478  1.360442029  2.41959173
#>  [2,]  0.75951402  0.63750620 -0.40213595  0.46266350  0.948918556  1.80133467
#>  [3,]  0.72729611  0.43829410 -0.31175145  0.48731162  1.064569890  1.71052377
#>  [4,]  0.28282493  0.13192850 -0.77608979 -0.10931576  0.206410088  1.00566051
#>  [5,]  0.86175454  0.50687354 -0.55701424  0.26850153  0.754827587  1.62323707
#>  [6,]  0.45025489  0.51006186 -0.57804143  0.23366524  0.597706237  1.57673941
#>  [7,]  0.74782294  0.31278427 -0.55138727  0.37576402  0.955170461  1.69587663
#>  [8,]  0.72249570  0.34318191 -0.66742995  0.16532620  0.631608122  1.40136989
#>  [9,]  0.56742679  0.35648481 -0.69353862  0.19652228  0.669437564  1.49049998
#> [10,]  0.70830429  0.52315628 -0.19747444  0.45815069  0.908316196  1.48093731
#> [11,]  0.69470616  0.20617007 -0.70877798  0.08545062  0.802651705  1.35390142
#> [12,]  0.48009838  0.16207417 -0.60452008  0.20608902  0.747972192  1.34357418
#> [13,]  0.97317242  0.80293258 -0.37039051  0.61637442  1.080853524  2.01474339
#> [14,]  1.01811669  0.68850159 -0.42883194  0.45646185  0.999322300  1.81150827
#> [15,]  0.16643882 -0.56542370 -1.11200360 -0.38387113  0.225551374  0.55070805
#> [16,] -0.26272917 -0.69661491 -1.38659994 -0.58718638 -0.048719918  0.42067567
#> [17,]  1.09165563  0.86104929 -0.13464999  0.71369779  1.338004103  2.14476654
#> [18,]  1.07287205  0.87650737 -0.28257742  0.69354093  1.219489173  2.17012126
#> [19,]  0.87025783  0.25820603 -0.72279553  0.11930408  0.968064506  1.62746913
#> [20,]  0.65647728  0.59384214 -0.23882809  0.60131870  1.112931668  1.58428883
#> [21,]  0.79934580  0.30100784 -0.71318727  0.22117629  0.961886945  1.58197619
#> [22,]  0.96871597  0.37095784 -0.53977798  0.39246994  1.124770263  1.78123430
#> [23,]  0.67387891  0.22145734 -0.74892391  0.10242957  0.817089968  1.53670938
#> [24,]  0.58271435  0.36555518 -0.32442584  0.22685914  0.845063488  1.39358674
#> [25,]  0.68415227  0.40248888 -0.44525644  0.46418207  1.004761163  1.62977604
#> [26,]  0.56751668  0.58010226 -0.31939969  0.42882365  0.798943276  1.55728987
#> [27,]  0.83769605  0.70167028 -0.50713458  0.41695017  0.922523980  1.93516384
#> [28,]  0.49776313  0.03122457 -0.85308648 -0.06252406  0.354708918  1.05762442
#> [29,]  0.09514963 -0.67373001 -1.48791114 -0.55886368 -0.014801890  0.56276801
#> [30,]  0.72638059  0.57050804 -0.17070673  0.55916813  1.129559048  1.54650259
#> [31,]  1.07103870  1.05390111 -0.15582476  0.76945802  1.205641840  2.13992086
#> [32,]  0.94086492  0.65087500 -0.22005331  0.72645570  1.300029168  1.98948951
#> [33,]  0.93371089  0.18190983 -0.64127262  0.15275177  0.969413099  1.71084031
#> [34,]  0.75664669  0.68399674 -0.52004814  0.44989655  0.833271370  1.92865394
#> [35,] -0.22513474 -1.23458153 -1.67614826 -0.95543925 -0.328815895 -0.04413418
#> [36,] -0.15920969 -0.33366361 -1.02404225 -0.29258437  0.002660234  0.44750578
#> [37,]  0.82936000  0.55949875 -0.28376746  0.58837647  1.176321050  1.82532073
#> [38,]  0.18038841  0.21556270 -0.78001466  0.11949968  0.344606567  1.21439226
#> [39,]  0.70750667  0.43948455 -0.43704956  0.44718010  0.967915443  1.63262634
#> [40,]  0.97721712  0.49436373 -0.31791347  0.54353478  1.246306689  1.87909958
#>            [,7]      [,8]        [,9]        [,10]
#>  [1,] 2.1727035 1.2659911  1.36776942  0.467152009
#>  [2,] 1.7522249 0.9641305  1.06396084  0.062391911
#>  [3,] 1.6945612 0.9174255  0.76497567  0.086097029
#>  [4,] 1.2243621 0.7987310  0.89033024 -0.052160212
#>  [5,] 1.8098886 1.1539393  1.03007933  0.178560116
#>  [6,] 1.5490027 0.7778972  1.14542517 -0.075419878
#>  [7,] 1.7509454 1.0606641  0.82453167  0.106925922
#>  [8,] 1.6723059 1.0800673  0.87287739  0.141078491
#>  [9,] 1.7361531 1.0930038  0.82964592 -0.044750339
#> [10,] 1.5394263 0.8653731  0.75107478  0.249814930
#> [11,] 1.8036597 1.1698974  0.45530211  0.162177563
#> [12,] 1.5085489 0.7866961  0.59123216 -0.014039486
#> [13,] 1.9563860 1.1444212  1.14939498  0.181758962
#> [14,] 1.9232762 1.1784350  1.03014929  0.196409003
#> [15,] 1.2248102 0.9351812  0.05099882  0.082338481
#> [16,] 1.0800504 0.3219133 -0.02446328 -0.574722694
#> [17,] 1.9856120 1.0977771  1.09515904  0.303728083
#> [18,] 2.0447569 1.1454819  1.18425905  0.225766381
#> [19,] 2.0419124 1.2811419  0.60100796  0.155305287
#> [20,] 1.5662646 0.9003293  0.73854172  0.331959418
#> [21,] 1.8037685 1.1220060  0.72223519  0.060555721
#> [22,] 2.0226680 1.3061910  0.68378827  0.330445759
#> [23,] 1.7097945 0.9759442  0.68704744  0.097075273
#> [24,] 1.4987583 0.8349312  0.76365920  0.071708518
#> [25,] 1.7002758 0.9499925  0.73884486  0.157835668
#> [26,] 1.4970002 0.8564409  0.83817835  0.106760251
#> [27,] 1.9179830 1.0540271  1.10914965  0.019311181
#> [28,] 1.5040766 1.0793682  0.61904318  0.007295298
#> [29,] 1.3891720 0.9608530  0.20817725 -0.170329340
#> [30,] 1.5762314 0.9372766  0.76059670  0.344330423
#> [31,] 1.9255839 1.1170943  1.31153583  0.365850064
#> [32,] 1.9056369 1.0996822  0.90840060  0.250564307
#> [33,] 1.9108827 1.1461251  0.95436267  0.125239962
#> [34,] 1.8170866 0.9694928  1.14828329  0.063689194
#> [35,] 1.0316909 0.6785760 -0.26822005 -0.449204793
#> [36,] 0.8862166 0.4454532  0.25077725 -0.371863750
#> [37,] 1.7571363 0.9543463  0.82814888  0.145586620
#> [38,] 1.3832590 0.8066259  0.88305844 -0.014021144
#> [39,] 1.6824127 0.9271002  0.77413168  0.133380139
#> [40,] 1.9891858 1.1707800  0.75139097  0.300788688
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2909736 -0.1055561
#>  [2,] 0.2727879 -0.1253604
#>  [3,] 0.2685352 -0.1184788
#>  [4,] 0.2460396 -0.1439961
#>  [5,] 0.2783737 -0.1301767
#>  [6,] 0.2607263 -0.1377576
#>  [7,] 0.2777526 -0.1254168
#>  [8,] 0.2719325 -0.1315401
#>  [9,] 0.2715294 -0.1354137
#> [10,] 0.2593590 -0.1124589
#> [11,] 0.2741056 -0.1246134
#> [12,] 0.2587334 -0.1266364
#> [13,] 0.2840284 -0.1199537
#> [14,] 0.2823421 -0.1220895
#> [15,] 0.2459757 -0.1351375
#> [16,] 0.2166229 -0.1535394
#> [17,] 0.2831400 -0.1115716
#> [18,] 0.2867050 -0.1168477
#> [19,] 0.2861087 -0.1283345
#> [20,] 0.2657522 -0.1052723
#> [21,] 0.2797214 -0.1281020
#> [22,] 0.2897571 -0.1169151
#> [23,] 0.2735801 -0.1298087
#> [24,] 0.2567026 -0.1251661
#> [25,] 0.2717316 -0.1175677
#> [26,] 0.2581355 -0.1184395
#> [27,] 0.2790640 -0.1294172
#> [28,] 0.2613972 -0.1392075
#> [29,] 0.2510557 -0.1562480
#> [30,] 0.2656494 -0.1069875
#> [31,] 0.2833426 -0.1112420
#> [32,] 0.2816311 -0.1108929
#> [33,] 0.2862538 -0.1367995
#> [34,] 0.2761790 -0.1279545
#> [35,] 0.2179868 -0.1611742
#> [36,] 0.2147002 -0.1415664
#> [37,] 0.2730545 -0.1144230
#> [38,] 0.2546982 -0.1379382
#> [39,] 0.2700517 -0.1187089
#> [40,] 0.2843304 -0.1130217
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5978536  0.5978536
#>  [2,]  0.4017573 -0.4017573
#>  [3,]  0.4044363 -0.4044363
#>  [4,]  0.4037087 -0.4037087
#>  [5,]  0.3992598 -0.3992598
#>  [6,]  0.4016766 -0.4016766
#>  [7,]  0.4005511 -0.4005511
#>  [8,]  0.4004783 -0.4004783
#>  [9,] -0.6003547  0.6003547
#> [10,] -0.5918982  0.5918982
#> [11,]  0.4016202 -0.4016202
#> [12,]  0.4048324 -0.4048324
#> [13,]  0.4003560 -0.4003560
#> [14,]  0.4002481 -0.4002481
#> [15,] -0.5941415  0.5941415
#> [16,]  0.4085018 -0.4085018
#> [17,]  0.4025836 -0.4025836
#> [18,]  0.4004591 -0.4004591
#> [19,]  0.3978472 -0.3978472
#> [20,] -0.5917065  0.5917065
#> [21,] -0.6005659  0.6005659
#> [22,] -0.6002897  0.6002897
#> [23,]  0.4004984 -0.4004984
#> [24,]  0.4056763 -0.4056763
#> [25,]  0.4038860 -0.4038860
#> [26,] -0.5930468  0.5930468
#> [27,]  0.3992763 -0.3992763
#> [28,] -0.5988329  0.5988329
#> [29,] -0.6004412  0.6004412
#> [30,] -0.5920960  0.5920960
#> [31,] -0.5973859  0.5973859
#> [32,]  0.4031098 -0.4031098
#> [33,]  0.3957864 -0.3957864
#> [34,] -0.5996804  0.5996804
#> [35,] -0.5936707  0.5936707
#> [36,]  0.4118636 -0.4118636
#> [37,]  0.4043247 -0.4043247
#> [38,] -0.5969172  0.5969172
#> [39,]  0.4040157 -0.4040157
#> [40,] -0.5980513  0.5980513
#> 
#> $L
#> [1] 0.6936015 0.6756555 0.6981082 0.6966312 0.7397603 0.6721005
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536