Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13          V14
#>  [1,] -0.1545714 -0.06439533 -0.18852350 -0.16075079  0.01038685 -0.133062682
#>  [2,] -0.3892866 -0.05080691 -0.17888652 -0.07989053 -0.01767078 -0.007584151
#>  [3,] -0.3488858 -0.13918180 -0.10497671 -0.06853657 -0.18415531 -0.147095773
#>  [4,] -0.1855831 -0.11269540 -0.09083972 -0.13393378 -0.12063233 -0.092476895
#>  [5,] -0.2236067  0.01385483  0.02685207 -0.08591268  0.06454609 -0.055160298
#>  [6,] -0.3026606 -0.19388727 -0.10021778 -0.06967962 -0.11826006 -0.137337377
#>  [7,] -0.3774119 -0.15286317 -0.14113952 -0.09046658 -0.14133927 -0.214306117
#>  [8,] -0.3572148 -0.07362630 -0.08114245 -0.02779923 -0.03132659 -0.054864654
#>  [9,] -0.3972124 -0.21779462 -0.14377327 -0.13681706 -0.05546496 -0.099624339
#> [10,] -0.2485202 -0.09506835 -0.17928763 -0.05348026 -0.22029122 -0.032833644
#>                V15           V16           V17          V18         V19
#>  [1,] -0.036139045 -0.1109801915 -0.1500941285 -0.041776265 -0.05109132
#>  [2,] -0.003588439 -0.0128483180 -0.0447180667  0.061114995 -0.01395140
#>  [3,] -0.024234571 -0.1454227757 -0.0874597606 -0.095638574 -0.01122667
#>  [4,] -0.089317642 -0.0675684393 -0.0247289637  0.020149990 -0.07205982
#>  [5,]  0.108776408  0.0004527771  0.1268741126  0.120423854  0.14672719
#>  [6,] -0.153791234 -0.0701587301 -0.0004650268 -0.041546271  0.05056329
#>  [7,] -0.241410845 -0.0895028671 -0.1973546628 -0.271718903 -0.04859819
#>  [8,] -0.040952314  0.0864512645  0.1374880924  0.199227595  0.14658313
#>  [9,] -0.109372998  0.0510150479 -0.1096697154 -0.038766547 -0.08031352
#> [10,] -0.042253167 -0.0438755059 -0.0727463813 -0.006159757 -0.05288342
#>               V2         V20         V21         V22         V23         V24
#>  [1,] -0.1810136  0.04337956 -0.07471319 -0.01565033 -0.05555635 -0.04236071
#>  [2,] -0.2801401  0.08468648  0.07828355  0.01977365  0.14756495  0.15462464
#>  [3,] -0.2898255  0.05235387  0.13671581  0.18526214  0.12153899  0.12668193
#>  [4,] -0.3263101 -0.05956669  0.03861458  0.07066997  0.16355632  0.06135182
#>  [5,] -0.2084159  0.07112671  0.12773076  0.12276671  0.09721555  0.12928772
#>  [6,] -0.3268316  0.10939133 -0.06420980 -0.04487157  0.01324935  0.01196888
#>  [7,] -0.3513805 -0.15594777 -0.20929222 -0.11717327 -0.13883555 -0.03431175
#>  [8,] -0.3455897  0.07849719  0.04659269  0.20488667  0.16672863  0.12215255
#>  [9,] -0.2971938  0.03356683 -0.07879218 -0.02566397 -0.03707073  0.01543175
#> [10,] -0.3164652 -0.02599082  0.05369718  0.11828435  0.02051499  0.04102668
#>               V25         V26        V27          V28         V29         V3
#>  [1,]  0.01053175 -0.02478386 0.09752902  0.001224718 -0.08270628 -0.1977073
#>  [2,]  0.07630252  0.06867482 0.08343964  0.089420091  0.02188784 -0.3272449
#>  [3,]  0.05472696  0.13483475 0.07229357  0.051447594  0.09729601 -0.2463474
#>  [4,]  0.12232413  0.06874982 0.18625038  0.240987026  0.08185280 -0.1892378
#>  [5,]  0.05389610  0.05622674 0.24063924  0.191331312  0.10899866 -0.2358803
#>  [6,] -0.01424890  0.13497346 0.11178051  0.117393146  0.06349652 -0.3193765
#>  [7,]  0.03330069  0.15105730 0.14234139  0.257121180  0.21572408 -0.2792789
#>  [8,]  0.25918407  0.16815067 0.22009298  0.185116499  0.21018089 -0.3460274
#>  [9,] -0.06576485 -0.02401497 0.09069438  0.008715803 -0.01340635 -0.3665347
#> [10,]  0.10186213  0.12314131 0.05935310 -0.065955436 -0.05741974 -0.3101469
#>               V30         V31         V32         V33          V34          V35
#>  [1,] -0.09383520 -0.07534395 -0.13582626  0.03350281 -0.133040199 -0.078284413
#>  [2,]  0.07155478  0.01377474  0.07506830  0.01846823  0.067631592  0.027120614
#>  [3,]  0.07624018  0.06318807 -0.08984833 -0.06523349  0.054263721  0.012142017
#>  [4,]  0.05794249 -0.02871896  0.04748839 -0.06783244 -0.001956408 -0.001713885
#>  [5,]  0.04572715  0.08425632  0.04858768 -0.01155169 -0.044951057  0.010918197
#>  [6,]  0.05459074 -0.05315877  0.09238804  0.05655175 -0.055413693 -0.061968143
#>  [7,]  0.18091336  0.07021261  0.11154122  0.05624925  0.069494910  0.083160089
#>  [8,]  0.23447792  0.10657431  0.15484889  0.08878765  0.058675023  0.040037853
#>  [9,] -0.04954391 -0.07292525 -0.04642981 -0.15233615 -0.187823637 -0.046623120
#> [10,] -0.11102703  0.04298126 -0.06355741 -0.10589618 -0.191904875 -0.050569396
#>               V36          V37          V38         V39         V4         V40
#>  [1,] -0.17953058 -0.089926410 -0.116394029 -0.04287553 -0.1311715 -0.01810860
#>  [2,]  0.02723188 -0.042524706 -0.097502004  0.01444346 -0.2484135 -0.06198955
#>  [3,] -0.06714720 -0.092660700 -0.109135521 -0.11470931 -0.2936453 -0.05131229
#>  [4,]  0.05467343 -0.015770625 -0.000684279 -0.05365195 -0.2288936  0.07858570
#>  [5,] -0.03494105  0.009967597 -0.069022838  0.05787554 -0.3017501  0.02005899
#>  [6,]  0.04125387  0.009845422  0.016516785 -0.12228932 -0.2320811 -0.04803182
#>  [7,]  0.09285119  0.030889503  0.027403346 -0.05615101 -0.2236479 -0.10196206
#>  [8,]  0.12165405 -0.026571055  0.030834907  0.10282698 -0.3262323  0.02380688
#>  [9,] -0.18277111 -0.048998703 -0.131074123 -0.13110397 -0.2872936 -0.06691330
#> [10,] -0.15438053 -0.030959338 -0.026906015 -0.14315911 -0.2229124 -0.20822260
#>                V41          V42          V43         V44          V45
#>  [1,] -0.087246399 -0.062745546 -0.037083628 -0.07836179 -0.229669538
#>  [2,] -0.121037378 -0.007908742 -0.115733765 -0.04719570 -0.049066830
#>  [3,] -0.028918201  0.031784462 -0.107078910 -0.06042275 -0.019500126
#>  [4,] -0.114505780 -0.063841761  0.024234583 -0.15394317 -0.014553398
#>  [5,]  0.006902996  0.029431994 -0.117271251 -0.10729906 -0.146953465
#>  [6,]  0.033814319 -0.186574189 -0.032273181 -0.02928754 -0.169068145
#>  [7,]  0.042099791 -0.073418414 -0.144848954 -0.15130712 -0.007400798
#>  [8,] -0.056942403  0.012302393 -0.006071172 -0.03687158 -0.152816041
#>  [9,] -0.035717172 -0.053964111 -0.227049109 -0.14045050 -0.078577936
#> [10,] -0.118706766 -0.230844606 -0.116024192 -0.16230642 -0.173754570
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.21142889 -0.1888719 -0.1181188 -0.1941638 -0.3014877 -0.2541217
#>  [2,] -0.13122979 -0.1272077 -0.2356776 -0.3151331 -0.2872857 -0.3827015
#>  [3,] -0.05808382 -0.2047527 -0.2860095 -0.3318016 -0.2264811 -0.4024230
#>  [4,] -0.17309818 -0.1215187 -0.1124650 -0.3225073 -0.2336786 -0.2940045
#>  [5,] -0.07543433 -0.1297099 -0.3075894 -0.2327037 -0.2876378 -0.3771134
#>  [6,] -0.19676988 -0.2593776 -0.1235164 -0.2138500 -0.1895064 -0.3730846
#>  [7,] -0.18399061 -0.2179490 -0.1792404 -0.1966894 -0.1975250 -0.3258862
#>  [8,] -0.09957887 -0.2069500 -0.2417276 -0.3632095 -0.2096238 -0.3581597
#>  [9,] -0.18331473 -0.2120362 -0.1895525 -0.3270733 -0.2659802 -0.3943377
#> [10,] -0.21086324 -0.1154487 -0.2440166 -0.2728193 -0.2167949 -0.2025312
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2148658 -0.1533860 -0.3188307 -0.2246976 -0.2846494 -0.2440137
#>  [2,] -0.2765222 -0.3947656 -0.3311440 -0.4284110 -0.3729192 -0.3083290
#>  [3,] -0.3396713 -0.4008416 -0.3148686 -0.3423884 -0.3806593 -0.3920883
#>  [4,] -0.3892044 -0.2928858 -0.2986200 -0.2909755 -0.3440091 -0.3929197
#>  [5,] -0.4265019 -0.4170824 -0.4169408 -0.3509910 -0.4417140 -0.4040122
#>  [6,] -0.3425841 -0.3075293 -0.2393140 -0.3083019 -0.3496414 -0.3067086
#>  [7,] -0.3905405 -0.2863968 -0.2623601 -0.4167314 -0.3236779 -0.2843334
#>  [8,] -0.2919902 -0.4171714 -0.4692852 -0.4652400 -0.3740818 -0.4422954
#>  [9,] -0.3282975 -0.2519782 -0.2832479 -0.3481754 -0.2532326 -0.2740509
#> [10,] -0.3148610 -0.2249665 -0.2352655 -0.2459924 -0.2734971 -0.2687656
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.1893862 -0.1606064 -0.2095112 -0.1293661 -0.1886409 -0.1263302
#>  [2,] -0.3359101 -0.3084722 -0.4595984 -0.2122364 -0.4365795 -0.2494525
#>  [3,] -0.3190883 -0.4630925 -0.3014186 -0.1278372 -0.3740686 -0.2668691
#>  [4,] -0.2684510 -0.4297148 -0.2590397 -0.2227211 -0.2362720 -0.1175870
#>  [5,] -0.4567811 -0.3594847 -0.2976167 -0.2540574 -0.3080807 -0.2268563
#>  [6,] -0.2781454 -0.2655533 -0.2616820 -0.2533631 -0.2847783 -0.1160955
#>  [7,] -0.4400360 -0.3541522 -0.3548120 -0.2455104 -0.2827163 -0.2687114
#>  [8,] -0.3149059 -0.3607895 -0.3981912 -0.2381861 -0.4455762 -0.2496927
#>  [9,] -0.3561172 -0.2762596 -0.4136369 -0.1407173 -0.3084293 -0.1780355
#> [10,] -0.2733730 -0.4084538 -0.4361319 -0.2055171 -0.3854330 -0.1626160
#>               V8          V9
#>  [1,] -0.2470877 -0.20622060
#>  [2,] -0.1863968 -0.08148924
#>  [3,] -0.2024818 -0.10595058
#>  [4,] -0.1163614 -0.14552554
#>  [5,] -0.0703623 -0.13259161
#>  [6,] -0.1905860 -0.10951807
#>  [7,] -0.2518465 -0.24526954
#>  [8,] -0.1880898 -0.10492240
#>  [9,] -0.2540977 -0.08202378
#> [10,] -0.2483592 -0.22068214
#> 
#> $W[[2]]
#>             [,1]          [,2]         [,3]         [,4]        [,5]
#> [1,] -0.06275309 -0.0008756129 -0.068213358 -0.088400554 -0.02279833
#> [2,] -0.03306566 -0.0553345074 -0.002815307  0.006089541  0.03626193
#>             [,6]        [,7]        [,8]        [,9]       [,10]
#> [1,] -0.09101308 -0.06933391  0.08788094 -0.08834033 0.002035349
#> [2,]  0.05912034  0.15010248 -0.07653591 -0.02703544 0.053712259
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.136256e-05  6.763068e-05  8.174945e-05  7.962771e-05  0.0001106285
#>  [2,] -5.765211e-05 -2.932448e-04 -5.028174e-04 -5.402420e-04 -0.0003439271
#>  [3,]  6.714130e-05  4.439435e-04  5.734390e-04  5.940127e-04  0.0006362701
#>  [4,]  9.211541e-05  5.710771e-04  7.567692e-04  7.903485e-04  0.0008359786
#>  [5,]  7.390328e-05  4.631341e-04  5.854671e-04  5.936403e-04  0.0006364284
#>  [6,]  1.263778e-04  8.109852e-04  1.136435e-03  1.187839e-03  0.0011221492
#>  [7,]  1.494643e-04  9.664365e-04  1.353208e-03  1.391242e-03  0.0012428827
#>  [8,] -8.112863e-05 -3.271260e-04 -5.242338e-04 -5.362472e-04 -0.0004621444
#>  [9,]  3.394066e-05  2.202308e-04  3.081380e-04  3.157941e-04  0.0003073762
#> [10,]  2.862320e-05  1.874609e-04  2.552058e-04  2.611853e-04  0.0002737600
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0001121749  0.0001291364  0.0002171160  2.530100e-04  2.814704e-04
#>  [2,] -0.0001145503  0.0001634161  0.0003265922  5.409923e-04  5.431496e-04
#>  [3,]  0.0005683751  0.0005297481  0.0007596653  8.219516e-04  9.134083e-04
#>  [4,]  0.0007385398  0.0006638714  0.0009303798  9.818053e-04  1.082189e-03
#>  [5,]  0.0005734824  0.0005262301  0.0007099245  7.590327e-04  8.393316e-04
#>  [6,]  0.0008599554  0.0006384688  0.0009327016  9.106905e-04  1.049566e-03
#>  [7,]  0.0008663506  0.0005458439  0.0009207495  8.523878e-04  9.547366e-04
#>  [8,] -0.0003378821 -0.0001734678 -0.0001432568 -2.449823e-05 -3.080488e-05
#>  [9,]  0.0002461944  0.0002173839  0.0003418980  3.616545e-04  4.161729e-04
#> [10,]  0.0002392403  0.0002305847  0.0003621176  3.953856e-04  4.489172e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  2.673080e-04  2.182958e-05  0.0002515094  2.252532e-04  0.0002995399
#>  [2,] -4.891673e-05 -9.781650e-05 -0.0002930694 -3.218071e-05  0.0001029664
#>  [3,]  1.101161e-03  1.247263e-04  0.0011374386  9.793488e-04  0.0011865030
#>  [4,]  1.369396e-03  1.701415e-04  0.0014280771  1.182509e-03  0.0014099581
#>  [5,]  1.038865e-03  1.312782e-04  0.0011050528  9.737425e-04  0.0011420469
#>  [6,]  1.651100e-03  2.392344e-04  0.0018049957  1.380731e-03  0.0016777555
#>  [7,]  1.725511e-03  2.972461e-04  0.0019371478  1.424766e-03  0.0018452888
#>  [8,] -4.498780e-04 -1.355114e-04 -0.0006190981 -3.733453e-04 -0.0003008957
#>  [9,]  5.379664e-04  6.512544e-05  0.0005513255  4.391330e-04  0.0005644956
#> [10,]  5.235177e-04  5.537813e-05  0.0005246748  4.398607e-04  0.0005618533
#>                 V23          V24          V25          V26           V27
#>  [1,]  3.517239e-04 0.0003710282 0.0004032823 0.0004134254  4.138684e-04
#>  [2,]  4.773596e-04 0.0008297863 0.0009743505 0.0008622675  5.627035e-04
#>  [3,]  1.168129e-03 0.0010595872 0.0010841597 0.0011661976  1.326315e-03
#>  [4,]  1.347111e-03 0.0011845538 0.0011995536 0.0012954646  1.526842e-03
#>  [5,]  1.102842e-03 0.0010132624 0.0010451825 0.0011261191  1.281743e-03
#>  [6,]  1.446458e-03 0.0010582507 0.0010101893 0.0011812206  1.613447e-03
#>  [7,]  1.574091e-03 0.0010642296 0.0009902546 0.0012324062  1.830088e-03
#>  [8,] -5.228884e-05 0.0001747240 0.0002592127 0.0002099410 -8.126305e-06
#>  [9,]  5.408969e-04 0.0004493027 0.0004439344 0.0004780055  5.709367e-04
#> [10,]  5.771596e-04 0.0005263407 0.0005379913 0.0005703547  6.325578e-04
#>                 V28          V29            V3          V30          V31
#>  [1,]  3.780213e-04 3.536595e-04  2.878926e-05 3.702168e-04 0.0003399037
#>  [2,]  4.491115e-04 5.378341e-04 -8.509772e-05 6.220773e-04 0.0009730298
#>  [3,]  1.254085e-03 1.128079e-03  1.407320e-04 1.138827e-03 0.0008872770
#>  [4,]  1.450886e-03 1.284063e-03  1.894280e-04 1.303524e-03 0.0009838687
#>  [5,]  1.228870e-03 1.099863e-03  1.463487e-04 1.102798e-03 0.0008619272
#>  [6,]  1.552555e-03 1.291731e-03  2.595889e-04 1.275353e-03 0.0006994417
#>  [7,]  1.769406e-03 1.410285e-03  3.237067e-04 1.438862e-03 0.0006952337
#>  [8,] -7.588826e-05 7.329937e-06 -1.387616e-04 4.329230e-06 0.0002663087
#>  [9,]  5.287995e-04 4.568064e-04  7.483906e-05 4.628234e-04 0.0003183470
#> [10,]  5.868541e-04 5.251160e-04  6.591360e-05 5.320216e-04 0.0004175578
#>                V32           V33           V34           V35           V36
#>  [1,] 2.843211e-04  2.363517e-04  2.733832e-04  3.461231e-04  3.259560e-04
#>  [2,] 6.946201e-04  1.219807e-03  1.335098e-03  1.555435e-03  1.846656e-03
#>  [3,] 7.751949e-04  3.741478e-04  4.355213e-04  5.742005e-04  3.998913e-04
#>  [4,] 8.876844e-04  3.166306e-04  3.843545e-04  5.397567e-04  2.766968e-04
#>  [5,] 7.676333e-04  3.733435e-04  4.077886e-04  5.092243e-04  3.397470e-04
#>  [6,] 6.805602e-04 -2.597428e-04 -2.208805e-04 -7.960015e-05 -5.029823e-04
#>  [7,] 6.572028e-04 -5.354007e-04 -4.763824e-04 -2.916173e-04 -8.048847e-04
#>  [8,] 9.820867e-05  5.711868e-04  6.021490e-04  6.961728e-04  9.686299e-04
#>  [9,] 2.854734e-04  5.637133e-05  8.964258e-05  1.694596e-04  6.474133e-05
#> [10,] 3.588826e-04  1.818310e-04  2.250868e-04  3.172582e-04  2.369848e-04
#>                 V37           V38           V39            V4          V40
#>  [1,]  0.0002608664  2.068749e-04  0.0001528123  2.266964e-05 0.0001789000
#>  [2,]  0.0012509257  3.281818e-04  0.0001245190 -1.536519e-04 0.0005953809
#>  [3,]  0.0004561352  6.700529e-04  0.0005849400  1.526750e-04 0.0004666488
#>  [4,]  0.0004087872  7.771356e-04  0.0006924126  2.119595e-04 0.0004935930
#>  [5,]  0.0004183926  6.424468e-04  0.0005812662  1.711721e-04 0.0004427392
#>  [6,] -0.0000667743  7.794867e-04  0.0007317696  3.049032e-04 0.0003174856
#>  [7,] -0.0002655774  8.546462e-04  0.0008150068  3.863597e-04 0.0002940717
#>  [8,]  0.0005866219 -4.805651e-05 -0.0001301546 -2.031613e-04 0.0002340825
#>  [9,]  0.0001324183  2.776674e-04  0.0002214733  7.878669e-05 0.0001650551
#> [10,]  0.0002433142  3.100564e-04  0.0002449412  6.445541e-05 0.0002193684
#>                V41           V42           V43           V44           V45
#>  [1,] 1.801762e-04  1.374607e-04  0.0001284891  0.0001166029  0.0001124917
#>  [2,] 3.303610e-04  8.427317e-05 -0.0001448689 -0.0003934521 -0.0006815181
#>  [3,] 5.415262e-04  5.078188e-04  0.0005546604  0.0006219342  0.0007155173
#>  [4,] 6.364421e-04  6.292561e-04  0.0006930896  0.0007911799  0.0009494018
#>  [5,] 5.208033e-04  5.083133e-04  0.0005462150  0.0006227899  0.0007297327
#>  [6,] 6.009993e-04  6.516350e-04  0.0008740657  0.0011399925  0.0014524691
#>  [7,] 7.046901e-04  7.213531e-04  0.0010287265  0.0013973527  0.0018396549
#>  [8,] 1.711828e-05 -1.385907e-04 -0.0002596807 -0.0004090114 -0.0006869614
#>  [9,] 2.234997e-04  1.955039e-04  0.0002462581  0.0003026224  0.0003644617
#> [10,] 2.540083e-04  2.151198e-04  0.0002416517  0.0002681912  0.0003043212
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0001024932  7.703003e-05  5.003169e-05  2.520232e-05  2.218017e-05
#>  [2,] -0.0006144385 -3.620676e-04 -2.516804e-04 -1.605027e-04 -1.405008e-04
#>  [3,]  0.0006341992  4.512145e-04  3.008028e-04  1.673132e-04  1.550207e-04
#>  [4,]  0.0008523862  5.976372e-04  3.973204e-04  2.225293e-04  2.096943e-04
#>  [5,]  0.0006435455  4.519870e-04  3.038514e-04  1.695472e-04  1.689774e-04
#>  [6,]  0.0013018662  8.912550e-04  5.932525e-04  3.413215e-04  2.953769e-04
#>  [7,]  0.0016607334  1.122767e-03  7.356125e-04  4.266320e-04  3.739971e-04
#>  [8,] -0.0006481020 -4.029610e-04 -2.692116e-04 -1.588004e-04 -1.744797e-04
#>  [9,]  0.0003296719  2.357077e-04  1.545673e-04  8.807013e-05  7.477858e-05
#> [10,]  0.0002747174  2.009483e-04  1.329004e-04  7.366352e-05  6.257876e-05
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.562467e-05  8.637665e-06  8.062355e-06  7.152261e-06  7.173018e-06
#>  [2,] -3.451183e-05 -5.268440e-05 -4.963925e-05 -1.980065e-05 -1.404054e-05
#>  [3,]  7.250700e-05  5.607296e-05  5.028281e-05  3.275602e-05  3.145216e-05
#>  [4,]  9.423058e-05  7.557600e-05  6.872656e-05  4.333367e-05  4.116421e-05
#>  [5,]  7.135107e-05  5.636976e-05  5.006225e-05  3.147952e-05  3.037343e-05
#>  [6,]  1.283700e-04  1.126270e-04  1.028283e-04  5.962395e-05  5.478232e-05
#>  [7,]  1.582338e-04  1.398309e-04  1.255385e-04  6.983927e-05  6.414892e-05
#>  [8,] -5.151383e-05 -5.655228e-05 -5.504020e-05 -2.655696e-05 -2.169680e-05
#>  [9,]  3.631510e-05  2.898167e-05  2.692331e-05  1.720500e-05  1.609284e-05
#> [10,]  3.301518e-05  2.410069e-05  2.239286e-05  1.577081e-05  1.518984e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  6.950092e-06  4.701594e-06  2.865413e-06  3.394017e-06  3.835345e-06
#>  [2,] -8.389904e-06 -1.356526e-05  1.817805e-05 -1.868249e-05 -1.047568e-05
#>  [3,]  2.863797e-05  2.388218e-05  4.674370e-06  2.187638e-05  1.928307e-05
#>  [4,]  3.828742e-05  3.177271e-05  4.161088e-06  3.102589e-05  2.591771e-05
#>  [5,]  2.824232e-05  2.392202e-05  5.920890e-06  2.279820e-05  1.934898e-05
#>  [6,]  4.697558e-05  4.216256e-05 -5.658866e-06  4.140541e-05  3.375085e-05
#>  [7,]  5.454680e-05  4.775231e-05 -9.183763e-06  4.621643e-05  4.153236e-05
#>  [8,] -2.018388e-05 -1.916070e-05  7.759488e-06 -2.383718e-05 -1.664197e-05
#>  [9,]  1.294385e-05  1.147966e-05 -4.145724e-07  1.038449e-05  9.071935e-06
#> [10,]  1.270451e-05  1.055728e-05  1.586187e-06  9.088264e-06  8.577838e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.185065e-05  4.008808e-06  4.593408e-05  2.519319e-05  5.074314e-05
#>  [2,]  1.005630e-05 -1.024217e-05  8.052248e-05  2.664389e-05 -2.822811e-04
#>  [3,]  1.348063e-04  1.902051e-05  1.579558e-04  1.275277e-04  3.606894e-04
#>  [4,]  1.610557e-04  2.530739e-05  1.833996e-04  1.459218e-04  4.663155e-04
#>  [5,]  1.378419e-04  1.947721e-05  1.615298e-04  1.466806e-04  3.870493e-04
#>  [6,]  1.779973e-04  3.328798e-05  1.710596e-04  1.404497e-04  6.735553e-04
#>  [7,]  2.097537e-04  4.036654e-05  1.820614e-04  1.589405e-04  8.273294e-04
#>  [8,] -5.225641e-05 -1.777877e-05 -2.381712e-05 -3.976540e-05 -3.076172e-04
#>  [9,]  5.991420e-05  9.531307e-06  6.637225e-05  4.216229e-05  1.737918e-04
#> [10,]  5.955010e-05  8.688789e-06  7.237914e-05  4.455564e-05  1.457382e-04
#> 
#> $vW[[2]]
#>              [,1]         [,2]         [,3]         [,4]        [,5]
#> [1,]  0.007515401  0.006902909  0.005933046  0.006715371 -0.00155641
#> [2,] -0.007515401 -0.006902909 -0.005933046 -0.006715371  0.00155641
#>             [,6]       [,7]         [,8]         [,9]        [,10]
#> [1,]  0.01107501  0.0202919 -0.002128139  0.007912645  0.008332465
#> [2,] -0.01107501 -0.0202919  0.002128139 -0.007912645 -0.008332465
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.006647038 0.505379316 0.384243031 0.546962332 0.521839482 0.297909260
#>  [7] 0.363843862 0.961996267 0.270664538 0.103238250
#> 
#> $B[[2]]
#> [1] -0.02373836 -0.14012304
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0005930337 0.0012705107 0.0017177761 0.0019620939 0.0016493585
#>  [6] 0.0017325129 0.0017227060 0.0001884105 0.0007021285 0.0008322043
#> 
#> $vB[[2]]
#> [1] -0.006682798  0.006682798
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [2,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#>  [3,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [4,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#>  [5,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#>  [6,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#>  [7,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#>  [8,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#>  [9,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [10,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [11,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [12,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [13,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [14,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [15,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [16,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [17,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [18,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [19,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [20,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [21,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [22,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [23,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [24,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [25,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [26,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [27,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [28,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [29,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [30,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [31,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [32,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [33,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [34,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [35,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [36,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#> [37,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [38,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [39,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [40,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [2,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#>  [3,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [4,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#>  [5,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#>  [6,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#>  [7,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#>  [8,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#>  [9,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [10,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [11,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [12,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [13,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [14,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [15,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [16,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [17,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [18,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [19,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [20,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [21,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [22,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [23,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [24,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [25,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [26,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [27,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [28,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [29,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [30,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [31,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [32,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [33,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [34,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [35,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [36,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#> [37,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [38,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [39,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [40,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [2,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#>  [3,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [4,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#>  [5,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#>  [6,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#>  [7,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#>  [8,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#>  [9,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [10,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [11,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [12,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [13,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [14,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [15,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [16,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [17,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [18,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [19,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [20,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [21,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [22,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [23,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [24,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [25,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [26,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [27,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [28,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [29,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [30,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [31,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [32,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [33,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [34,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [35,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [36,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#> [37,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [38,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [39,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [40,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [2,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#>  [3,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [4,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#>  [5,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#>  [6,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#>  [7,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#>  [8,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#>  [9,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [10,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [11,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [12,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [13,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [14,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [15,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [16,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [17,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [18,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [19,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [20,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [21,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [22,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [23,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [24,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [25,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [26,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [27,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [28,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [29,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [30,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [31,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [32,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [33,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [34,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [35,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [36,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#> [37,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [38,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [39,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [40,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [2,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#>  [3,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [4,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#>  [5,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#>  [6,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#>  [7,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#>  [8,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#>  [9,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [10,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [11,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [12,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [13,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [14,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [15,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [16,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [17,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [18,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [19,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [20,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [21,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [22,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [23,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [24,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [25,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [26,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [27,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [28,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [29,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [30,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [31,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [32,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [33,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [34,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [35,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [36,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#> [37,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [38,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [39,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [40,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [2,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#>  [3,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [4,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#>  [5,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#>  [6,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#>  [7,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#>  [8,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#>  [9,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [10,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [11,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [12,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [13,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [14,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [15,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [16,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [17,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [18,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [19,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [20,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [21,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [22,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [23,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [24,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [25,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [26,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [27,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [28,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [29,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [30,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [31,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [32,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [33,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [34,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [35,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [36,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#> [37,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [38,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [39,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [40,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3763939 0.6876942 0.6549451 0.7211947 0.7576196 0.5768554 0.6164835
#>  [2,] 0.1680999 0.5933929 0.4649655 0.5703457 0.7325120 0.3968635 0.3190526
#>  [3,] 0.2249212 0.6042214 0.4794780 0.5237826 0.7589370 0.4264225 0.2420386
#>  [4,] 0.3079595 0.6895747 0.6181704 0.6966171 0.7688584 0.5529429 0.6064556
#>  [5,] 0.1746677 0.6192339 0.5073975 0.5920176 0.7397786 0.4414157 0.3621550
#>  [6,] 0.1727548 0.6255535 0.5492572 0.6210017 0.7221674 0.4213487 0.5015670
#>  [7,] 0.2791471 0.6588557 0.6241147 0.6676470 0.7544106 0.4784110 0.4188729
#>  [8,] 0.2955956 0.6476466 0.5828688 0.6753631 0.7843570 0.5356630 0.4545673
#>  [9,] 0.2391387 0.6541963 0.4940427 0.5636039 0.7768895 0.4435979 0.2545591
#> [10,] 0.2423159 0.7023279 0.6428162 0.7433733 0.8122863 0.5707252 0.5889211
#> [11,] 0.3014964 0.7162289 0.6586967 0.7473949 0.8085648 0.5914030 0.6322486
#> [12,] 0.2677268 0.7570506 0.6937006 0.7803540 0.8354506 0.6217398 0.6715757
#> [13,] 0.1833854 0.6211271 0.5195521 0.6536438 0.7665728 0.4290588 0.3665582
#> [14,] 0.2518973 0.6647182 0.6051314 0.7227193 0.7682957 0.5394314 0.5921833
#> [15,] 0.2773988 0.7234241 0.6648086 0.7406413 0.8097881 0.5983100 0.6471068
#> [16,] 0.2414944 0.5731592 0.4126167 0.5122193 0.7493357 0.4199848 0.2714311
#> [17,] 0.2096803 0.5802276 0.4771612 0.5426638 0.7625252 0.3964299 0.1936153
#> [18,] 0.3016560 0.7276662 0.6876892 0.7426729 0.8331792 0.5742667 0.5262901
#> [19,] 0.1765068 0.6143727 0.4956049 0.6554701 0.7193385 0.4837138 0.6185427
#> [20,] 0.2650491 0.6556765 0.6056709 0.6895133 0.7709700 0.5437011 0.5540859
#> [21,] 0.1871773 0.5941259 0.5196252 0.6414851 0.7230899 0.4146912 0.4403388
#> [22,] 0.3368990 0.7189330 0.6686903 0.7478898 0.8179164 0.5946822 0.5733079
#> [23,] 0.2483589 0.6154215 0.5172685 0.6500790 0.6907280 0.5244908 0.6590873
#> [24,] 0.1776152 0.6088473 0.4811971 0.6487470 0.7316055 0.4992305 0.6124423
#> [25,] 0.2282134 0.6023616 0.4964284 0.6376949 0.6864417 0.5032665 0.6393695
#> [26,] 0.2511956 0.6835513 0.5916184 0.6876818 0.8185169 0.5382381 0.4415627
#> [27,] 0.2651635 0.6740419 0.5498838 0.6560757 0.8039403 0.5505645 0.4546498
#> [28,] 0.3013693 0.7090583 0.6657818 0.7385542 0.8106551 0.5750440 0.5969834
#> [29,] 0.1679619 0.5587154 0.4850866 0.5819403 0.7103918 0.3979092 0.4300874
#> [30,] 0.2911087 0.6468642 0.5893181 0.6754979 0.7764355 0.5122735 0.4640525
#> [31,] 0.3079128 0.6893849 0.6087584 0.7168188 0.7872546 0.5682934 0.5798017
#> [32,] 0.2628052 0.7137102 0.6466176 0.7222230 0.8472027 0.5434484 0.4122952
#> [33,] 0.3119786 0.6993979 0.6677310 0.7359854 0.8003985 0.5718502 0.5721729
#> [34,] 0.2368993 0.7100956 0.6079635 0.7343009 0.8162278 0.5773631 0.6141112
#> [35,] 0.2389820 0.6760100 0.5921918 0.6592056 0.8229459 0.4872520 0.3159203
#> [36,] 0.2629389 0.6968199 0.5920334 0.6896780 0.7912887 0.5147474 0.5394596
#> [37,] 0.3133123 0.7421201 0.6899432 0.7156694 0.8360384 0.5602151 0.4242091
#> [38,] 0.2962494 0.7193960 0.6586477 0.7478725 0.8221137 0.5820257 0.5964204
#> [39,] 0.2246227 0.6827296 0.5776922 0.6608974 0.7918488 0.5039161 0.4149517
#> [40,] 0.2043968 0.4996899 0.4030453 0.5327944 0.6395892 0.3734812 0.3464935
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8865618 0.4289787 0.4420992
#>  [2,] 0.8963932 0.1779948 0.2034020
#>  [3,] 0.8673324 0.2698260 0.2846662
#>  [4,] 0.9062705 0.3435738 0.3460995
#>  [5,] 0.8982418 0.1927067 0.2226043
#>  [6,] 0.8922749 0.2093406 0.2156013
#>  [7,] 0.8746623 0.3338680 0.3774315
#>  [8,] 0.9003071 0.3424971 0.3572060
#>  [9,] 0.8893493 0.2828548 0.3252825
#> [10,] 0.9380123 0.2851898 0.3198821
#> [11,] 0.9317369 0.3521352 0.4104926
#> [12,] 0.9528998 0.3178448 0.3805973
#> [13,] 0.9020201 0.2288587 0.2777618
#> [14,] 0.9102302 0.2921055 0.3110011
#> [15,] 0.9356331 0.3334160 0.3571874
#> [16,] 0.8458428 0.2771444 0.2841830
#> [17,] 0.8637468 0.2422283 0.2844025
#> [18,] 0.9337609 0.3535324 0.4138340
#> [19,] 0.9078936 0.1982667 0.1918662
#> [20,] 0.9059666 0.3207477 0.3357206
#> [21,] 0.8756934 0.2172986 0.2211590
#> [22,] 0.9265863 0.3970248 0.4467913
#> [23,] 0.8678897 0.2763165 0.2397631
#> [24,] 0.9112127 0.2097732 0.2011516
#> [25,] 0.8667757 0.2526706 0.2149396
#> [26,] 0.9297861 0.3014855 0.3209299
#> [27,] 0.9082348 0.3245647 0.3309197
#> [28,] 0.9287822 0.3494263 0.3940135
#> [29,] 0.8677293 0.2104276 0.2024074
#> [30,] 0.8908820 0.3581445 0.3874847
#> [31,] 0.9130248 0.3507881 0.3804470
#> [32,] 0.9411783 0.3237192 0.3865587
#> [33,] 0.9202577 0.3665250 0.4304360
#> [34,] 0.9449631 0.2880016 0.3425764
#> [35,] 0.9236694 0.2950492 0.3474714
#> [36,] 0.9225743 0.2778302 0.2880095
#> [37,] 0.9354144 0.3734366 0.4552325
#> [38,] 0.9355820 0.3467935 0.3936909
#> [39,] 0.9221169 0.2315514 0.2784924
#> [40,] 0.7912571 0.2270107 0.2213724
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4816932 0.5183068
#>  [2,] 0.5112214 0.4887786
#>  [3,] 0.5111154 0.4888846
#>  [4,] 0.4866006 0.5133994
#>  [5,] 0.5069057 0.4930943
#>  [6,] 0.5001890 0.4998110
#>  [7,] 0.4962535 0.5037465
#>  [8,] 0.4936103 0.5063897
#>  [9,] 0.5099044 0.4900956
#> [10,] 0.4874851 0.5125149
#> [11,] 0.4828690 0.5171310
#> [12,] 0.4809075 0.5190925
#> [13,] 0.5045344 0.4954656
#> [14,] 0.4881565 0.5118435
#> [15,] 0.4830659 0.5169341
#> [16,] 0.5097559 0.4902441
#> [17,] 0.5136502 0.4863498
#> [18,] 0.4877064 0.5122936
#> [19,] 0.4937815 0.5062185
#> [20,] 0.4894316 0.5105684
#> [21,] 0.5017146 0.4982854
#> [22,] 0.4839912 0.5160088
#> [23,] 0.4877763 0.5122237
#> [24,] 0.4935186 0.5064814
#> [25,] 0.4902247 0.5097753
#> [26,] 0.4959510 0.5040490
#> [27,] 0.4949709 0.5050291
#> [28,] 0.4849537 0.5150463
#> [29,] 0.5039655 0.4960345
#> [30,] 0.4931514 0.5068486
#> [31,] 0.4866165 0.5133835
#> [32,] 0.4953121 0.5046879
#> [33,] 0.4852921 0.5147079
#> [34,] 0.4869359 0.5130641
#> [35,] 0.5032126 0.4967874
#> [36,] 0.4931546 0.5068454
#> [37,] 0.4928724 0.5071276
#> [38,] 0.4849842 0.5150158
#> [39,] 0.5003255 0.4996745
#> [40,] 0.5071215 0.4928785
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]         [,2]        [,3]       [,4]      [,5]         [,6]
#>  [1,] -0.5048826  0.789361404  0.64084799 0.95039533 1.1396732  0.309877602
#>  [2,] -1.5991537  0.378009405 -0.14036821 0.28326176 1.0074051 -0.418551052
#>  [3,] -1.2372148  0.423085822 -0.08213402 0.09520228 1.1468603 -0.296462625
#>  [4,] -0.8096762  0.798131857  0.48178975 0.83124025 1.2018764  0.212568566
#>  [5,] -1.5529006  0.486297911  0.02959232 0.37231250 1.0448182 -0.235418569
#>  [6,] -1.5662279  0.513188107  0.19766997 0.49380208 0.9552383 -0.317239245
#>  [7,] -0.9486960  0.658198895  0.50705007 0.69756195 1.1222759 -0.086409796
#>  [8,] -0.8683603  0.608710348  0.33456127 0.73254306 1.2912401  0.142894589
#>  [9,] -1.1574072  0.637536033 -0.02383017 0.25580129 1.2476309 -0.226572765
#> [10,] -1.1400245  0.858407917  0.58760824 1.06357595 1.4649347  0.284810666
#> [11,] -0.8401821  0.925831788  0.65749157 1.08476621 1.4407112  0.369768286
#> [12,] -1.0061867  1.136576868  0.81747722 1.26773057 1.6247603  0.496939546
#> [13,] -1.4935772  0.494335115  0.07824828 0.63509477 1.1890593 -0.285692325
#> [14,] -1.0885185  0.684391854  0.42689252 0.95799039 1.1987129  0.158053794
#> [15,] -0.9574014  0.961510192  0.68479746 1.04930450 1.4486341  0.398428340
#> [16,] -1.1445041  0.294752257 -0.35315862 0.04888681 1.0950725 -0.322835898
#> [17,] -1.3268534  0.323707854 -0.09141886 0.17107105 1.1665739 -0.420363078
#> [18,] -0.8394245  0.982814135  0.78933812 1.05990783 1.6083285  0.299280795
#> [19,] -1.5401956  0.465730198 -0.01758090 0.64317190 0.9411828 -0.065167913
#> [20,] -1.0198889  0.644085749  0.42915108 0.79784496 1.2137967  0.175251664
#> [21,] -1.4684565  0.381048235  0.07854111 0.58181576 0.9598407 -0.344605267
#> [22,] -0.6771441  0.939175205  0.70226759 1.08738911 1.5022943  0.383355528
#> [23,] -1.1073842  0.470159619  0.06910152 0.61938639 0.8035249  0.098041745
#> [24,] -1.5325893  0.442469353 -0.07524720 0.61353622 1.0027834 -0.003077925
#> [25,] -1.2184270  0.415314856 -0.01428672 0.56537332 0.7835362  0.013066293
#> [26,] -1.0922461  0.770140793  0.37065976 0.78930386 1.5063317  0.153251514
#> [27,] -1.0193014  0.726523266  0.20020130 0.64585458 1.4111057  0.202951612
#> [28,] -0.8407856  0.890814679  0.68916812 1.03846742 1.4542727  0.302461087
#> [29,] -1.6001410  0.235950389 -0.05967136 0.33074359 0.8972879 -0.414184400
#> [30,] -0.8900056  0.605283620  0.36114711 0.73315797 1.2450134  0.049103967
#> [31,] -0.8098952  0.797245129  0.44209613 0.92873593 1.3084558  0.274891594
#> [32,] -1.0314393  0.913472516  0.60420445 0.95551555 1.7128279  0.174232985
#> [33,] -0.7908854  0.844432488  0.69794063 1.02520605 1.3887868  0.289404133
#> [34,] -1.1697546  0.895848461  0.43875990 1.01655458 1.4909964  0.311957943
#> [35,] -1.1582687  0.735495271  0.37303349 0.65975624 1.5364352 -0.051003093
#> [36,] -1.0307493  0.832200168  0.37237759 0.79861457 1.3327109  0.059006516
#> [37,] -0.7846794  1.057016985  0.79985362 0.92308092 1.6290425  0.242035134
#> [38,] -0.8652226  0.941467331  0.65727363 1.08729763 1.5307344  0.331094676
#> [39,] -1.2389277  0.766344267  0.31330691 0.66729611 1.3361058  0.015664852
#> [40,] -1.3590375 -0.001240438 -0.39279221 0.13136615 0.5735815 -0.517311235
#>               [,7]     [,8]       [,9]      [,10]
#>  [1,]  0.474648942 2.056092 -0.2860193 -0.2326469
#>  [2,] -0.758129267 2.157776 -1.5299926 -1.3651660
#>  [3,] -1.141535503 1.877575 -0.9955054 -0.9214318
#>  [4,]  0.432437614 2.268925 -0.6474085 -0.6362289
#>  [5,] -0.566023169 2.177840 -1.4325179 -1.2505535
#>  [6,]  0.006268191 2.114192 -1.3289048 -1.2914868
#>  [7,] -0.327402079 1.942826 -0.6907422 -0.5004646
#>  [8,] -0.182233383 2.200642 -0.6521860 -0.5875117
#>  [9,] -1.074442934 2.084112 -0.9303448 -0.7296001
#> [10,]  0.359507210 2.716827 -0.9188622 -0.7543137
#> [11,]  0.541875321 2.613681 -0.6096668 -0.3619293
#> [12,]  0.715320416 3.007232 -0.7636941 -0.4870136
#> [13,] -0.547011020 2.219874 -1.2147670 -0.9555913
#> [14,]  0.372998215 2.316449 -0.8851801 -0.7954430
#> [15,]  0.606345920 2.676623 -0.6927751 -0.5875928
#> [16,] -0.987374104 1.702361 -0.9586710 -0.9238062
#> [17,] -1.426687976 1.846765 -1.1405016 -0.9227273
#> [18,]  0.105257465 2.645950 -0.6035479 -0.3481382
#> [19,]  0.483367273 2.288183 -1.3971630 -1.4379296
#> [20,]  0.217193567 2.265353 -0.7503377 -0.6824235
#> [21,] -0.239787093 1.952265 -1.2814790 -1.2589252
#> [22,]  0.295360215 2.535397 -0.4178773 -0.2136439
#> [23,]  0.659229519 1.882427 -0.9628071 -1.1539789
#> [24,]  0.457590160 2.328533 -1.3262929 -1.3791122
#> [25,]  0.572628610 1.872746 -1.0844195 -1.2954034
#> [26,] -0.234822166 2.583408 -0.8402338 -0.7495017
#> [27,] -0.181900718 2.292270 -0.7328723 -0.7040285
#> [28,]  0.392911433 2.568131 -0.6215621 -0.4304726
#> [29,] -0.281494526 1.881029 -1.3223499 -1.3713157
#> [30,] -0.144038580 2.099782 -0.5834268 -0.4578977
#> [31,]  0.321959277 2.351140 -0.6155767 -0.4876515
#> [32,] -0.354485381 2.772622 -0.7367320 -0.4618008
#> [33,]  0.290721888 2.445854 -0.5471537 -0.2800725
#> [34,]  0.464626703 2.843142 -0.9051094 -0.6518338
#> [35,] -0.772584305 2.493280 -0.8709861 -0.6301726
#> [36,]  0.158167416 2.477849 -0.9552501 -0.9050711
#> [37,] -0.305517911 2.672998 -0.5175019 -0.1795510
#> [38,]  0.390572121 2.675776 -0.6331636 -0.4318237
#> [39,] -0.343532273 2.471462 -1.1995715 -0.9519523
#> [40,] -0.634488259 1.332520 -1.2252682 -1.2576867
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] -0.1954594 -0.1221995
#>  [2,] -0.1122273 -0.1571204
#>  [3,] -0.1206659 -0.1651350
#>  [4,] -0.1776766 -0.1240661
#>  [5,] -0.1235252 -0.1511498
#>  [6,] -0.1350821 -0.1358380
#>  [7,] -0.1603698 -0.1453835
#>  [8,] -0.1651685 -0.1396082
#>  [9,] -0.1264729 -0.1660956
#> [10,] -0.1738014 -0.1237314
#> [11,] -0.1869743 -0.1184233
#> [12,] -0.1904811 -0.1140740
#> [13,] -0.1316787 -0.1498167
#> [14,] -0.1701409 -0.1227581
#> [15,] -0.1854525 -0.1176903
#> [16,] -0.1197893 -0.1588178
#> [17,] -0.1147723 -0.1693866
#> [18,] -0.1822715 -0.1330871
#> [19,] -0.1435001 -0.1186248
#> [20,] -0.1691941 -0.1269142
#> [21,] -0.1343878 -0.1412460
#> [22,] -0.1908514 -0.1267943
#> [23,] -0.1621340 -0.1132293
#> [24,] -0.1439326 -0.1180058
#> [25,] -0.1544735 -0.1153675
#> [26,] -0.1590005 -0.1428042
#> [27,] -0.1595213 -0.1394041
#> [28,] -0.1839497 -0.1237464
#> [29,] -0.1246960 -0.1405584
#> [30,] -0.1654934 -0.1380975
#> [31,] -0.1787521 -0.1252055
#> [32,] -0.1655176 -0.1467656
#> [33,] -0.1845123 -0.1256635
#> [34,] -0.1718005 -0.1195322
#> [35,] -0.1456830 -0.1585337
#> [36,] -0.1610755 -0.1336923
#> [37,] -0.1757328 -0.1472207
#> [38,] -0.1839395 -0.1238583
#> [39,] -0.1452573 -0.1465591
#> [40,] -0.1178584 -0.1463464
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4816932  0.4816932
#>  [2,]  0.4887786 -0.4887786
#>  [3,]  0.4888846 -0.4888846
#>  [4,] -0.4866006  0.4866006
#>  [5,] -0.5069057  0.5069057
#>  [6,]  0.4998110 -0.4998110
#>  [7,]  0.5037465 -0.5037465
#>  [8,]  0.5063897 -0.5063897
#>  [9,]  0.4900956 -0.4900956
#> [10,]  0.5125149 -0.5125149
#> [11,]  0.5171310 -0.5171310
#> [12,] -0.4809075  0.4809075
#> [13,] -0.5045344  0.5045344
#> [14,]  0.5118435 -0.5118435
#> [15,] -0.4830659  0.4830659
#> [16,]  0.4902441 -0.4902441
#> [17,] -0.5136502  0.5136502
#> [18,]  0.5122936 -0.5122936
#> [19,]  0.5062185 -0.5062185
#> [20,]  0.5105684 -0.5105684
#> [21,]  0.4982854 -0.4982854
#> [22,] -0.4839912  0.4839912
#> [23,] -0.4877763  0.4877763
#> [24,]  0.5064814 -0.5064814
#> [25,] -0.4902247  0.4902247
#> [26,]  0.5040490 -0.5040490
#> [27,] -0.4949709  0.4949709
#> [28,]  0.5150463 -0.5150463
#> [29,]  0.4960345 -0.4960345
#> [30,] -0.4931514  0.4931514
#> [31,] -0.4866165  0.4866165
#> [32,]  0.5046879 -0.5046879
#> [33,]  0.5147079 -0.5147079
#> [34,] -0.4869359  0.4869359
#> [35,]  0.4967874 -0.4967874
#> [36,] -0.4931546  0.4931546
#> [37,] -0.4928724  0.4928724
#> [38,] -0.4849842  0.4849842
#> [39,] -0.5003255  0.5003255
#> [40,]  0.4928785 -0.4928785
#> 
#> $L
#> [1] 0.6919978 0.7038474 0.6935672 0.6993209 0.7477295 0.6893709
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4492754