Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10           V11          V12          V13
#>  [1,] -0.2824951 -0.14938612 -0.0176825532 -0.050804822 -0.114371001
#>  [2,] -0.3770087 -0.03976830 -0.0541246510 -0.178771242 -0.079280426
#>  [3,] -0.3960208 -0.03120573  0.0004835521 -0.069527540 -0.097970595
#>  [4,] -0.2466362 -0.11536800 -0.0847939394 -0.130145677 -0.066303523
#>  [5,] -0.2608934 -0.17155534 -0.1860973623 -0.078515117 -0.034820121
#>  [6,] -0.2119384 -0.16369127 -0.0984888057 -0.151988469 -0.030713450
#>  [7,] -0.3813826 -0.09167225 -0.0293049641 -0.005793127  0.017793172
#>  [8,] -0.3518714 -0.23308816 -0.2385371496 -0.151755189 -0.160473814
#>  [9,] -0.1858444 -0.01779398 -0.0761568728 -0.013973356  0.008204922
#> [10,] -0.2494724 -0.01718642 -0.0428484493 -0.074632603 -0.112861487
#>               V14          V15          V16          V17          V18
#>  [1,] -0.11127979  0.037759901 -0.048100814  0.092926400  0.139300302
#>  [2,] -0.04607304  0.003789362 -0.032636400 -0.043147115  0.017347166
#>  [3,]  0.01003417 -0.027248549  0.064823714  0.070278942 -0.002765688
#>  [4,]  0.01346609 -0.033710552  0.021833661 -0.019324837 -0.099472735
#>  [5,] -0.14991220 -0.191963298 -0.054882803 -0.044798686 -0.135853006
#>  [6,] -0.09988458  0.033465552 -0.002921987  0.027382253 -0.012503293
#>  [7,] -0.12887129 -0.157468675 -0.004956298  0.037462955  0.051530561
#>  [8,] -0.04236372 -0.076840343 -0.034698135 -0.052196564 -0.122429697
#>  [9,]  0.02443561  0.017481020  0.081046070  0.011825757  0.215239057
#> [10,] -0.08624210  0.054346090 -0.027085141  0.009113168 -0.038546888
#>                V19         V2          V20         V21         V22         V23
#>  [1,]  0.022612284 -0.2834061  0.105636251  0.12446587  0.16482104 -0.02405100
#>  [2,]  0.009736247 -0.2331283  0.149722215  0.13795904  0.02183013  0.18144938
#>  [3,]  0.067106735 -0.3908464  0.169060392  0.10728469  0.18538687  0.20978221
#>  [4,] -0.099481512 -0.3758646  0.017899349  0.03625538 -0.03114982  0.09163502
#>  [5,] -0.013973947 -0.3340652  0.075519384  0.04078618  0.03930653  0.04147819
#>  [6,]  0.054640561 -0.2327312  0.023729321 -0.02075809 -0.02036264  0.02006371
#>  [7,]  0.164892669 -0.2781811  0.004907466  0.19456271  0.21617541  0.21547502
#>  [8,] -0.083737690 -0.2483052 -0.072250552 -0.12869966 -0.06648185 -0.05156417
#>  [9,]  0.126829598 -0.2829582  0.121700774  0.10448888  0.25187238  0.15399861
#> [10,]  0.073991331 -0.1723815  0.024816643  0.05250478 -0.10364258  0.06209377
#>               V24         V25         V26         V27         V28         V29
#>  [1,]  0.06454614  0.17545889  0.22755211  0.18526551  0.16839877 0.006832463
#>  [2,]  0.23313388  0.18987632  0.16929412  0.18550149  0.29286127 0.238286160
#>  [3,]  0.24428431  0.16485370  0.21353867  0.18237621  0.17189914 0.036385836
#>  [4,]  0.07598173  0.12489723  0.23991766  0.15830554  0.08818119 0.066359435
#>  [5,]  0.15583210  0.16669183  0.05080344  0.19670263  0.11453844 0.145156001
#>  [6,]  0.01044687  0.10924940  0.15234589  0.10811735  0.05459105 0.021999142
#>  [7,]  0.14908854  0.26988828  0.27153144  0.18485320  0.20320202 0.147110560
#>  [8,] -0.06178208 -0.04313006  0.13188543  0.07555986  0.05610603 0.201512561
#>  [9,]  0.18395547  0.29311196  0.33145124  0.23145525  0.17255338 0.082410349
#> [10,] -0.10022707 -0.12019016 -0.10706702 -0.03314454 -0.02905253 0.022436016
#>               V3          V30          V31          V32         V33         V34
#>  [1,] -0.3830970 -0.068804818 -0.105637442  0.005617243  0.04265541  0.02053491
#>  [2,] -0.3323459  0.096465111  0.003379230  0.004423099 -0.01060942 -0.02882562
#>  [3,] -0.3114642  0.117211671  0.057660895  0.099965951  0.04349569 -0.00985005
#>  [4,] -0.2481155  0.077828980  0.093236446  0.062599958  0.07266482 -0.10843553
#>  [5,] -0.2704551  0.034909852  0.084555721  0.006829312 -0.02306119 -0.10024302
#>  [6,] -0.1310638  0.009393734 -0.002842555  0.035047783 -0.03671950  0.02483642
#>  [7,] -0.2746341  0.238985430  0.114263708  0.022264323  0.03464891  0.02031400
#>  [8,] -0.2219056  0.152541880  0.067637049  0.042978273 -0.08001759  0.07394566
#>  [9,] -0.2918570  0.078512756  0.138394863  0.092440733  0.06302438 -0.02384926
#> [10,] -0.2088959 -0.067404036 -0.166735486 -0.193572770 -0.12581231  0.03947101
#>                V35         V36         V37         V38           V39
#>  [1,] -0.063177733  0.06089630 -0.11371421  0.02052504  0.0312595027
#>  [2,] -0.006803049 -0.03899894 -0.00797170  0.09228997 -0.0001096902
#>  [3,] -0.030150440 -0.08676200 -0.09542172 -0.06819816  0.0432192116
#>  [4,]  0.084464565  0.08441913 -0.04976025 -0.07863487 -0.0985544187
#>  [5,] -0.024569921 -0.10442904 -0.02716291 -0.10522635 -0.0927918362
#>  [6,]  0.009686777  0.02119274  0.01198994  0.02645487 -0.0398679391
#>  [7,]  0.091927478  0.03833757  0.03807313  0.02651865  0.0208558182
#>  [8,] -0.089437363  0.09986220  0.09808564  0.08969221 -0.0591248637
#>  [9,]  0.096949642 -0.03400323  0.12869865  0.01192271  0.0007793220
#> [10,] -0.112174638 -0.08854492  0.01688866 -0.13413405 -0.1284165118
#>                V4         V40          V41          V42         V43         V44
#>  [1,] -0.35552085 -0.09062187 -0.159488681  0.004833275 -0.20967902 -0.15921460
#>  [2,] -0.21987591 -0.06678872  0.010569072 -0.073341497 -0.14092981 -0.01021120
#>  [3,] -0.31441926  0.05573400 -0.010982224  0.060018829 -0.10985260 -0.02850118
#>  [4,] -0.31684247 -0.11868829 -0.079416893 -0.144314289 -0.11129123 -0.04243388
#>  [5,] -0.19414540 -0.09684811 -0.026350578 -0.125309174 -0.02045846 -0.09857889
#>  [6,] -0.06884247 -0.14074684  0.003602834 -0.132089286 -0.03190247 -0.09844551
#>  [7,] -0.34317589 -0.06083497 -0.036483743 -0.092738442 -0.01788933 -0.06678449
#>  [8,] -0.33755102 -0.05252802 -0.077911384 -0.064029465 -0.02264278 -0.16348771
#>  [9,] -0.23246867  0.01026119 -0.053430209 -0.049307786  0.03838342 -0.14000451
#> [10,] -0.30652436  0.02963540 -0.195731298 -0.133948891 -0.13513525 -0.16195273
#>                V45         V46         V47        V48        V49         V5
#>  [1,] -0.126040285 -0.27032488 -0.20037705 -0.3765373 -0.4019594 -0.3412931
#>  [2,] -0.000452308 -0.06877605 -0.24893458 -0.1633179 -0.2537402 -0.3132424
#>  [3,] -0.089743899 -0.20685720 -0.16168358 -0.2097842 -0.3821245 -0.2524226
#>  [4,] -0.206911728 -0.06166413 -0.15479698 -0.2740486 -0.2315185 -0.2065568
#>  [5,] -0.132740046 -0.18779113 -0.09476248 -0.1039345 -0.1687672 -0.1432338
#>  [6,] -0.161884940 -0.08629012 -0.25026922 -0.2450913 -0.2752247 -0.1140461
#>  [7,] -0.204801546 -0.12531506 -0.13696741 -0.1881021 -0.2045627 -0.1940015
#>  [8,] -0.082214506 -0.06136772 -0.21585736 -0.2511222 -0.3360818 -0.2742927
#>  [9,] -0.058222892 -0.15526422 -0.16377545 -0.2077416 -0.2484658 -0.1698127
#> [10,] -0.080601920 -0.10247043 -0.14589434 -0.1105514 -0.3179010 -0.2997193
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.4397982 -0.4374066 -0.3212178 -0.3236214 -0.4617275 -0.3216165
#>  [2,] -0.2907708 -0.2330030 -0.2847574 -0.3200072 -0.3323666 -0.3343731
#>  [3,] -0.3207307 -0.4381690 -0.3451868 -0.3422300 -0.4152407 -0.3769515
#>  [4,] -0.4403110 -0.2942552 -0.3807226 -0.3011475 -0.2425294 -0.2857329
#>  [5,] -0.1903960 -0.2696591 -0.3424357 -0.2421913 -0.2350359 -0.3474304
#>  [6,] -0.1468923 -0.2030490 -0.1535299 -0.2923358 -0.2197363 -0.2980629
#>  [7,] -0.4213815 -0.3146166 -0.4508576 -0.3282589 -0.4279063 -0.4325336
#>  [8,] -0.2428593 -0.3076705 -0.3615333 -0.2914277 -0.3084864 -0.3836915
#>  [9,] -0.2106007 -0.3677358 -0.2243959 -0.3663090 -0.3540127 -0.2928019
#> [10,] -0.2415899 -0.3271884 -0.2064938 -0.3399442 -0.2073237 -0.2779673
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4634083 -0.4014357 -0.4115383 -0.3016425 -0.2492873 -0.3231807
#>  [2,] -0.2543813 -0.3663384 -0.2949391 -0.4319866 -0.1568889 -0.4369348
#>  [3,] -0.4216217 -0.4753232 -0.3733920 -0.4461072 -0.2197333 -0.4520494
#>  [4,] -0.2864124 -0.4281499 -0.4116360 -0.2775764 -0.1883744 -0.2989815
#>  [5,] -0.2479546 -0.2519384 -0.2631192 -0.1973793 -0.2315669 -0.3396694
#>  [6,] -0.1992739 -0.2909234 -0.3175234 -0.1554814 -0.1009175 -0.2192056
#>  [7,] -0.4642960 -0.3820163 -0.3377626 -0.4083387 -0.1306266 -0.3931335
#>  [8,] -0.3071611 -0.3281184 -0.3294116 -0.3205478 -0.1450735 -0.3734657
#>  [9,] -0.2824737 -0.2308451 -0.2137238 -0.2631945 -0.1767943 -0.2566267
#> [10,] -0.2316652 -0.2173556 -0.3375418 -0.2823096 -0.1726107 -0.2768292
#>                V7          V8          V9
#>  [1,] -0.21193156 -0.12420211 -0.24351918
#>  [2,] -0.07431716 -0.10407013 -0.03977458
#>  [3,] -0.26563344 -0.22629917 -0.08103713
#>  [4,] -0.24507675 -0.14193251 -0.22400793
#>  [5,] -0.15744577 -0.19934759 -0.02974092
#>  [6,] -0.13698543 -0.07873537 -0.09465689
#>  [7,] -0.15043492 -0.17279451 -0.17520638
#>  [8,] -0.16079225 -0.14161671 -0.22195969
#>  [9,] -0.11981220 -0.12143098 -0.01102663
#> [10,] -0.21602215 -0.23681622 -0.15440694
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]      [,4]       [,5]        [,6]
#> [1,] -0.06494897 0.12353113 0.112485368 0.0339620 0.06910518  0.03803241
#> [2,]  0.04690216 0.06505552 0.006823414 0.1032086 0.10259444 -0.04073546
#>             [,7]       [,8]       [,9]        [,10]
#> [1,]  0.02589058 -0.1474921 -0.0037004 -0.032265398
#> [2,] -0.09105992  0.1207283 -0.1147382 -0.001120163
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.841990e-04  1.263799e-03  1.593717e-03  1.464376e-03  1.615719e-03
#>  [2,]  3.383061e-05  1.843614e-04  2.173827e-04  2.012738e-04  2.058949e-04
#>  [3,] -5.592148e-06 -7.871601e-05 -1.069645e-04 -6.903949e-05 -1.013566e-04
#>  [4,]  1.159682e-04  8.144865e-04  1.045482e-03  1.013153e-03  1.109080e-03
#>  [5,]  9.689091e-05  6.668119e-04  8.621835e-04  8.455358e-04  8.986644e-04
#>  [6,] -8.707765e-05 -5.942095e-04 -8.023747e-04 -7.777513e-04 -7.956465e-04
#>  [7,] -2.397051e-05 -1.707110e-04 -2.217930e-04 -1.670030e-04 -2.147670e-04
#>  [8,]  2.699595e-04  2.026075e-03  2.652052e-03  2.587341e-03  2.730391e-03
#>  [9,] -2.418545e-06 -5.172510e-05 -7.029306e-05 -3.535733e-05 -6.313257e-05
#> [10,]  2.648201e-05  1.973669e-04  2.652443e-04  2.748127e-04  3.036124e-04
#>                 V14           V15          V16          V17          V18
#>  [1,]  1.358614e-03  8.729946e-04 6.187275e-04 5.317427e-04 5.748887e-04
#>  [2,]  1.923458e-04  1.754442e-04 1.757815e-04 1.711482e-04 1.724669e-04
#>  [3,] -7.896751e-05  5.151330e-06 7.899598e-05 1.172486e-04 1.162162e-04
#>  [4,]  9.387166e-04  6.200479e-04 4.767305e-04 4.719335e-04 5.472910e-04
#>  [5,]  7.343051e-04  4.564967e-04 3.215110e-04 3.165479e-04 3.846754e-04
#>  [6,] -5.602270e-04 -1.933092e-04 5.027302e-05 1.165022e-04 4.286662e-05
#>  [7,] -1.828713e-04 -7.794020e-05 1.846327e-05 6.856208e-05 6.759077e-05
#>  [8,]  2.128435e-03  1.114906e-03 6.362741e-04 6.105621e-04 8.054812e-04
#>  [9,] -5.151249e-05  8.164252e-06 6.031157e-05 8.520979e-05 9.116920e-05
#> [10,]  2.580124e-04  1.747419e-04 1.481500e-04 1.728205e-04 2.094798e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  1.551292e-03  2.001460e-04  2.373866e-03  0.0026439056  0.0024115444
#>  [2,]  2.897401e-04  3.843664e-05  3.628329e-04  0.0003533942  0.0003330879
#>  [3,]  1.630294e-05 -3.977827e-06 -1.187540e-04 -0.0001800957 -0.0001161888
#>  [4,]  1.172158e-03  1.257330e-04  1.648660e-03  0.0018247311  0.0017384406
#>  [5,]  9.167115e-04  1.044066e-04  1.291530e-03  0.0014143337  0.0013568308
#>  [6,] -5.287904e-04 -8.766971e-05 -9.582434e-04 -0.0010981119 -0.0009816074
#>  [7,] -1.043896e-04 -2.332159e-05 -3.050271e-04 -0.0003730963 -0.0002744564
#>  [8,]  2.447701e-03  2.859914e-04  3.723443e-03  0.0042433044  0.0040075673
#>  [9,]  2.841308e-05 -7.575278e-07 -7.528895e-05 -0.0001205393 -0.0000712971
#> [10,]  3.539219e-04  2.855568e-05  4.483431e-04  0.0004900297  0.0004738826
#>                 V23           V24           V25           V26           V27
#>  [1,]  1.949367e-03  1.738197e-03  1.865763e-03  2.077849e-03  2.132937e-03
#>  [2,]  3.597314e-04  3.588035e-04  3.640899e-04  3.580687e-04  3.506318e-04
#>  [3,]  4.598762e-05  1.081009e-04  8.845017e-05  3.200266e-05  2.149585e-05
#>  [4,]  1.471811e-03  1.320243e-03  1.374021e-03  1.497985e-03  1.527931e-03
#>  [5,]  1.172442e-03  1.039992e-03  1.073393e-03  1.150482e-03  1.176461e-03
#>  [6,] -6.893144e-04 -5.083765e-04 -5.359793e-04 -6.038549e-04 -6.496080e-04
#>  [7,] -7.725828e-05 -4.242256e-06 -2.987724e-05 -8.558183e-05 -9.139705e-05
#>  [8,]  3.175197e-03  2.671662e-03  2.813593e-03  3.166991e-03  3.312985e-03
#>  [9,]  5.947826e-05  1.073637e-04  8.990623e-05  5.004839e-05  4.557390e-05
#> [10,]  3.962048e-04  3.459216e-04  3.577108e-04  3.885275e-04  3.953606e-04
#>                 V28           V29            V3           V30           V31
#>  [1,]  1.932008e-03  0.0013760593  2.072588e-04  1.017340e-03  8.487700e-05
#>  [2,]  3.985567e-04  0.0003824882  3.993615e-05  3.317876e-04  2.427570e-04
#>  [3,]  1.134615e-04  0.0002149982 -5.071984e-06  2.037939e-04  3.268366e-04
#>  [4,]  1.351963e-03  0.0010026352  1.329114e-04  7.277039e-04  1.262554e-04
#>  [5,]  1.072421e-03  0.0008156842  1.093215e-04  5.701474e-04  8.743308e-05
#>  [6,] -5.439743e-04 -0.0002668697 -9.149503e-05 -3.837457e-05  4.603112e-04
#>  [7,] -1.584602e-05  0.0001012990 -2.477053e-05  1.022544e-04  2.792595e-04
#>  [8,]  2.776088e-03  0.0017918345  3.030239e-04  1.075077e-03 -5.389286e-04
#>  [9,]  1.158895e-04  0.0001883628 -1.102819e-06  1.708610e-04  2.546980e-04
#> [10,]  3.195594e-04  0.0002248655  3.194459e-05  1.426653e-04 -9.633261e-06
#>                 V32          V33           V34           V35           V36
#>  [1,]  2.087868e-04 2.904918e-04  8.940403e-05 -2.043606e-04 -8.294460e-04
#>  [2,]  2.222935e-04 2.038740e-04  1.658672e-04  9.387386e-05  2.949694e-05
#>  [3,]  2.695480e-04 2.241074e-04  2.162757e-04  1.979679e-04  2.934123e-04
#>  [4,]  2.055679e-04 2.829354e-04  1.571615e-04 -1.409036e-05 -3.813841e-04
#>  [5,]  1.494413e-04 1.900619e-04  5.116411e-05 -1.283199e-04 -4.291624e-04
#>  [6,]  3.246359e-04 2.649380e-04  4.351808e-04  6.164285e-04  9.493283e-04
#>  [7,]  2.158221e-04 1.691050e-04  1.695041e-04  1.830854e-04  3.500111e-04
#>  [8,] -2.556995e-04 1.680014e-06 -4.002393e-04 -8.467958e-04 -1.762675e-03
#>  [9,]  2.135664e-04 1.849623e-04  1.766420e-04  1.632835e-04  2.410143e-04
#> [10,]  2.310395e-05 6.112060e-05  3.564587e-05  5.798793e-06 -8.084932e-05
#>                 V37           V38           V39            V4           V40
#>  [1,] -8.676700e-04 -3.582495e-04 -1.325962e-04  3.240653e-04 -2.251425e-04
#>  [2,]  3.146030e-06  9.608595e-05  1.513447e-04  5.998568e-05  1.431848e-04
#>  [3,]  2.835120e-04  2.468744e-04  2.647226e-04 -1.367832e-05  2.899053e-04
#>  [4,] -4.145956e-04 -1.267216e-04  1.181016e-05  1.993830e-04 -2.933795e-05
#>  [5,] -4.414626e-04 -1.424322e-04  2.617199e-05  1.675499e-04 -9.481692e-07
#>  [6,]  8.857573e-04  4.739406e-04  2.483814e-04 -1.506869e-04  3.098821e-04
#>  [7,]  3.556515e-04  2.665968e-04  2.675246e-04 -4.427772e-05  3.080685e-04
#>  [8,] -1.739887e-03 -9.315653e-04 -5.158306e-04  4.844824e-04 -5.813016e-04
#>  [9,]  2.362619e-04  2.150966e-04  2.338698e-04 -8.077950e-06  2.493852e-04
#> [10,] -8.070161e-05 -3.241169e-05 -1.004293e-05  4.639887e-05 -8.992450e-06
#>                 V41           V42           V43           V44           V45
#>  [1,]  3.765464e-04  1.201789e-03  1.273294e-03  1.429127e-03  1.839779e-03
#>  [2,]  1.963748e-04  2.513631e-04  2.304585e-04  1.936007e-04  2.339218e-04
#>  [3,]  1.958274e-04  5.957472e-05  7.034241e-06 -7.214063e-05 -1.338908e-04
#>  [4,]  3.081050e-04  8.254685e-04  8.711588e-04  9.607606e-04  1.167947e-03
#>  [5,]  2.872986e-04  7.117740e-04  7.407429e-04  7.875367e-04  9.662627e-04
#>  [6,] -3.611555e-05 -5.392081e-04 -6.150676e-04 -6.966319e-04 -9.255607e-04
#>  [7,]  1.678574e-04 -3.086242e-05 -8.916211e-05 -1.659260e-04 -2.606273e-04
#>  [8,]  4.036160e-04  1.820574e-03  2.009780e-03  2.373055e-03  2.963307e-03
#>  [9,]  1.755923e-04  7.437561e-05  3.292677e-05 -3.174964e-05 -7.858164e-05
#> [10,]  5.811821e-05  1.874056e-04  2.021244e-04  2.371238e-04  2.745802e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.195123e-03  7.067398e-04  5.761447e-04  2.653202e-04  3.247765e-04
#>  [2,]  2.164090e-04  1.607456e-04  1.111166e-04  6.406730e-05  7.261293e-05
#>  [3,] -1.823144e-05  3.310033e-05  2.500773e-06  1.788360e-05  1.183687e-05
#>  [4,]  7.325151e-04  4.409685e-04  3.591666e-04  1.675849e-04  2.157668e-04
#>  [5,]  6.390909e-04  4.041978e-04  3.188467e-04  1.565405e-04  1.866718e-04
#>  [6,] -5.808741e-04 -3.461784e-04 -2.849592e-04 -1.334123e-04 -1.435604e-04
#>  [7,] -1.134614e-04 -2.783405e-05 -4.508254e-05 -5.269361e-06 -1.385123e-05
#>  [8,]  1.755177e-03  1.025705e-03  8.668341e-04  3.886414e-04  5.112164e-04
#>  [9,]  4.014600e-06  3.653203e-05  8.309050e-06  1.776508e-05  1.225844e-05
#> [10,]  1.461504e-04  9.087705e-05  8.107802e-05  3.638270e-05  5.066254e-05
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.998684e-05  9.953916e-05  9.585648e-05  5.561255e-05  4.525477e-05
#>  [2,]  2.216161e-05  1.930245e-05  1.570537e-05  1.081186e-05  1.009218e-05
#>  [3,]  1.945729e-05 -9.353755e-07 -5.183443e-06 -8.128284e-08  6.862093e-07
#>  [4,]  1.865394e-05  6.300611e-05  6.037595e-05  3.744576e-05  2.939824e-05
#>  [5,]  2.288661e-05  5.437548e-05  4.935801e-05  3.092393e-05  2.431219e-05
#>  [6,] -4.679236e-06 -4.679901e-05 -4.228117e-05 -2.308548e-05 -1.758238e-05
#>  [7,]  1.401778e-05 -1.022563e-05 -1.366953e-05 -5.485464e-06 -3.974222e-06
#>  [8,]  7.626103e-06  1.426046e-04  1.416284e-04  8.096254e-05  6.048677e-05
#>  [9,]  1.634022e-05  2.783433e-07 -3.071043e-06  8.616601e-07  1.213218e-06
#> [10,]  1.910100e-07  1.357188e-05  1.382287e-05  8.834304e-06  6.576040e-06
#>                 V55           V56           V57           V58           V59
#>  [1,]  2.174304e-05  4.663721e-05  3.610410e-06  4.153091e-05  3.150656e-05
#>  [2,]  9.367545e-06  3.878960e-06  7.003233e-06  8.307106e-06  8.019890e-06
#>  [3,]  5.530256e-06 -6.136370e-06  6.507707e-06 -9.694627e-07  1.491128e-06
#>  [4,]  1.340160e-05  3.074565e-05  1.652851e-06  2.597974e-05  2.076793e-05
#>  [5,]  1.225703e-05  2.256345e-05  2.634973e-06  2.150882e-05  1.711548e-05
#>  [6,] -4.183829e-06 -1.966486e-05  4.993829e-06 -1.741615e-05 -1.095195e-05
#>  [7,]  2.230352e-06 -9.698632e-06  4.515782e-06 -5.958995e-06 -2.384495e-06
#>  [8,]  1.555262e-05  7.216694e-05 -1.128572e-05  5.263498e-05  3.988624e-05
#>  [9,]  4.626359e-06 -4.287210e-06  4.992658e-06 -3.011863e-07  1.449736e-06
#> [10,]  2.130557e-06  8.424465e-06 -1.426389e-06  5.938604e-06  4.618271e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  4.225705e-04  1.457961e-05  3.744322e-04  5.613956e-04  0.0013172096
#>  [2,]  6.086477e-05  4.434265e-06  7.842600e-05  1.146999e-04  0.0001723530
#>  [3,] -1.238141e-05  1.509031e-06  1.579426e-05 -2.587528e-07 -0.0001215177
#>  [4,]  2.983844e-04  9.878039e-06  2.687576e-04  3.680452e-04  0.0008247584
#>  [5,]  2.308632e-04  7.598869e-06  2.124709e-04  3.039438e-04  0.0006661085
#>  [6,] -1.609172e-04 -2.804329e-06 -1.177490e-04 -2.236068e-04 -0.0006147341
#>  [7,] -4.026207e-05 -5.315953e-07 -1.322921e-05 -4.838363e-05 -0.0002199324
#>  [8,]  6.976212e-04  1.378860e-05  5.621170e-04  8.287860e-04  0.0020876731
#>  [9,] -5.886659e-06  1.615900e-06  1.749433e-05  3.356951e-06 -0.0000890364
#> [10,]  8.083324e-05  2.042853e-06  7.267783e-05  8.469178e-05  0.0001935276
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]         [,6]
#> [1,] -0.01822746 -0.02804091 -0.0292522 -0.01083695 -0.01506459 -0.006558965
#> [2,]  0.01822746  0.02804091  0.0292522  0.01083695  0.01506459  0.006558965
#>             [,7]         [,8]        [,9]        [,10]
#> [1,] -0.02736696  0.005547436 -0.03117845 -0.002792206
#> [2,]  0.02736696 -0.005547436  0.03117845  0.002792206
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.60407894 0.51459242 0.63502718 0.30798215 0.26329921 0.09461924
#>  [7] 0.52701921 0.28166780 0.57041569 0.16030114
#> 
#> $B[[2]]
#> [1]  0.12531661 -0.08545287
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0017537597  0.0005074662  0.0002965853  0.0013610524  0.0010343431
#>  [6] -0.0001609919  0.0001521468  0.0021781268  0.0002690847  0.0003530676
#> 
#> $vB[[2]]
#> [1] -0.03412996  0.03412996
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#>  [2,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#>  [3,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#>  [4,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#>  [5,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#>  [6,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#>  [7,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#>  [8,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#>  [9,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [10,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [11,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [12,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [13,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [14,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [15,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [16,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [17,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [18,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [19,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [20,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [21,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [22,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [23,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [24,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [25,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [26,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [27,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [28,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [29,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [30,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [31,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [32,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [33,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [34,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [35,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [36,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [37,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [38,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [39,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [40,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#>  [2,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#>  [3,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#>  [4,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#>  [5,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#>  [6,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#>  [7,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#>  [8,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#>  [9,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [10,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [11,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [12,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [13,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [14,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [15,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [16,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [17,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [18,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [19,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [20,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [21,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [22,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [23,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [24,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [25,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [26,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [27,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [28,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [29,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [30,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [31,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [32,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [33,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [34,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [35,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [36,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [37,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [38,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [39,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [40,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#>  [2,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#>  [3,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#>  [4,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#>  [5,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#>  [6,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#>  [7,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#>  [8,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#>  [9,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [10,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [11,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [12,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [13,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [14,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [15,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [16,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [17,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [18,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [19,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [20,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [21,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [22,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [23,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [24,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [25,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [26,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [27,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [28,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [29,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [30,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [31,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [32,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [33,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [34,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [35,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [36,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [37,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [38,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [39,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [40,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#>  [2,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#>  [3,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#>  [4,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#>  [5,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#>  [6,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#>  [7,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#>  [8,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#>  [9,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [10,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [11,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [12,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [13,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [14,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [15,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [16,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [17,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [18,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [19,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [20,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [21,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [22,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [23,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [24,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [25,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [26,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [27,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [28,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [29,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [30,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [31,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [32,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [33,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [34,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [35,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [36,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [37,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [38,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [39,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [40,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#>  [2,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#>  [3,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#>  [4,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#>  [5,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#>  [6,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#>  [7,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#>  [8,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#>  [9,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [10,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [11,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [12,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [13,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [14,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [15,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [16,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [17,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [18,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [19,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [20,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [21,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [22,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [23,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [24,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [25,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [26,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [27,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [28,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [29,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [30,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [31,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [32,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [33,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [34,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [35,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [36,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [37,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [38,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [39,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [40,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#>  [2,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#>  [3,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#>  [4,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#>  [5,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#>  [6,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#>  [7,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#>  [8,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#>  [9,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [10,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [11,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [12,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [13,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [14,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [15,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [16,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [17,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [18,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [19,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [20,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [21,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [22,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [23,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [24,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [25,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [26,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [27,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [28,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [29,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [30,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [31,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [32,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [33,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [34,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [35,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [36,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [37,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [38,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [39,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [40,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5115296 0.7457001 0.7706263 0.5179646 0.4736060 0.4061366 0.8045578
#>  [2,] 0.7296898 0.7887701 0.8431223 0.5390260 0.5145557 0.4987614 0.8516199
#>  [3,] 0.7342407 0.8287184 0.8378365 0.6465222 0.6303103 0.5480065 0.8561939
#>  [4,] 0.6290533 0.7840776 0.7916098 0.5895805 0.5672406 0.4806427 0.8176537
#>  [5,] 0.7299687 0.8377291 0.8565817 0.6648936 0.6372062 0.5430242 0.8759331
#>  [6,] 0.6599979 0.7649700 0.7833808 0.5530105 0.5036865 0.4742794 0.7920196
#>  [7,] 0.7095612 0.8240826 0.8477442 0.6588951 0.6007364 0.5418486 0.8890871
#>  [8,] 0.7089894 0.7715201 0.8244232 0.5269146 0.4726501 0.4854984 0.8342493
#>  [9,] 0.4108250 0.6064100 0.5895464 0.3766557 0.3318612 0.3559049 0.6641551
#> [10,] 0.4827863 0.6909249 0.6677353 0.4137636 0.3381619 0.3840785 0.7159569
#> [11,] 0.6523914 0.7145423 0.7866471 0.4486495 0.4184741 0.4347600 0.7787383
#> [12,] 0.4952377 0.6836344 0.6419277 0.4080375 0.3462118 0.3912564 0.6947217
#> [13,] 0.7589901 0.8635349 0.8596238 0.6577229 0.6045269 0.5552841 0.9128363
#> [14,] 0.7436079 0.8528077 0.8471320 0.6381844 0.5962851 0.5490887 0.8989382
#> [15,] 0.7200967 0.8183569 0.8473106 0.5911658 0.5565793 0.4965868 0.8660636
#> [16,] 0.6715317 0.8143142 0.8397615 0.5515736 0.5453024 0.4440840 0.8554314
#> [17,] 0.7440484 0.8398597 0.8569934 0.6705420 0.6419594 0.5561395 0.8793838
#> [18,] 0.6799795 0.7906109 0.8298387 0.5687809 0.5471242 0.4872143 0.8615204
#> [19,] 0.6388606 0.7620279 0.7634868 0.5530510 0.4893606 0.4928027 0.8140560
#> [20,] 0.6434518 0.6539295 0.7211671 0.4673713 0.3940251 0.4525422 0.7138923
#> [21,] 0.7100455 0.7859970 0.8453996 0.5625681 0.4985410 0.4889807 0.8563213
#> [22,] 0.6943353 0.8187490 0.8235977 0.6262935 0.6099052 0.5151592 0.8507180
#> [23,] 0.7315788 0.7897216 0.8505770 0.6115802 0.5712242 0.5212299 0.8667648
#> [24,] 0.5432744 0.6731768 0.7042990 0.4492916 0.3832771 0.4332996 0.7687943
#> [25,] 0.5531224 0.6911434 0.6894060 0.5295273 0.4564212 0.4640430 0.7613924
#> [26,] 0.7491751 0.8350475 0.8515235 0.6836994 0.5969495 0.5725563 0.8962869
#> [27,] 0.7415016 0.8290202 0.8668371 0.6074095 0.5904969 0.5262361 0.8858912
#> [28,] 0.7269465 0.7665796 0.8212982 0.5650837 0.5287605 0.5068035 0.8501804
#> [29,] 0.7111753 0.7794597 0.8284751 0.5601123 0.5305569 0.5062089 0.8366492
#> [30,] 0.7157239 0.7971617 0.8193912 0.5787189 0.5348296 0.5220637 0.8376065
#> [31,] 0.5910621 0.7941837 0.7916766 0.5953048 0.5288550 0.4709843 0.8473033
#> [32,] 0.7360832 0.7981210 0.8252363 0.6434735 0.5817543 0.5638951 0.8517748
#> [33,] 0.7191863 0.8225157 0.8528550 0.6076794 0.5662448 0.4946034 0.8665997
#> [34,] 0.6930835 0.6927670 0.7823860 0.4968310 0.4269329 0.4793667 0.7818060
#> [35,] 0.5637050 0.7013800 0.7488185 0.4894884 0.4438297 0.4294272 0.7466226
#> [36,] 0.7049726 0.7763930 0.8296140 0.5515091 0.4990347 0.4985940 0.8436944
#> [37,] 0.6893656 0.8280528 0.8382630 0.6872629 0.6219100 0.5530389 0.8696256
#> [38,] 0.6883053 0.8090197 0.8184402 0.6396832 0.6128751 0.5297280 0.8371137
#> [39,] 0.7416200 0.8526610 0.8678193 0.6999725 0.6751478 0.5564565 0.8956443
#> [40,] 0.7790102 0.8717924 0.8805940 0.7110204 0.6675948 0.5895743 0.9163446
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4055493 0.8421446 0.1636919
#>  [2,] 0.3709651 0.8993007 0.3161758
#>  [3,] 0.5280394 0.8779157 0.3415678
#>  [4,] 0.4988142 0.8473319 0.2762467
#>  [5,] 0.5366001 0.8975296 0.3050876
#>  [6,] 0.4310141 0.8529866 0.2990447
#>  [7,] 0.5324003 0.8990897 0.2716271
#>  [8,] 0.3526120 0.8913988 0.2876437
#>  [9,] 0.4326766 0.7122373 0.1827006
#> [10,] 0.3618390 0.7995362 0.1816255
#> [11,] 0.2528191 0.8606854 0.2872745
#> [12,] 0.3663700 0.7631854 0.2223169
#> [13,] 0.5219333 0.9251288 0.2665829
#> [14,] 0.5306655 0.9131384 0.2765846
#> [15,] 0.3801069 0.9105008 0.2576767
#> [16,] 0.3783585 0.8897577 0.2539786
#> [17,] 0.5441154 0.9021678 0.3061499
#> [18,] 0.4129092 0.8889832 0.2635391
#> [19,] 0.4490839 0.8593893 0.2410083
#> [20,] 0.3187236 0.8260030 0.3310312
#> [21,] 0.3732741 0.9094865 0.2862698
#> [22,] 0.5091899 0.8763239 0.2962445
#> [23,] 0.4026985 0.9012929 0.2890788
#> [24,] 0.4494742 0.8197286 0.2074715
#> [25,] 0.5541689 0.7764061 0.2586818
#> [26,] 0.5593452 0.9187602 0.2509065
#> [27,] 0.4196440 0.9151283 0.2854579
#> [28,] 0.3994015 0.8868988 0.3365375
#> [29,] 0.4162797 0.8760463 0.3558943
#> [30,] 0.4724279 0.8782291 0.3203701
#> [31,] 0.5044977 0.8755826 0.1820780
#> [32,] 0.5417489 0.8844006 0.3062973
#> [33,] 0.4097258 0.9085759 0.2651198
#> [34,] 0.3410903 0.8550321 0.3497448
#> [35,] 0.4522245 0.8093990 0.2359409
#> [36,] 0.3794689 0.8941407 0.2731582
#> [37,] 0.5722950 0.8904710 0.2781854
#> [38,] 0.5546301 0.8573442 0.3242827
#> [39,] 0.5288887 0.9097668 0.2722567
#> [40,] 0.5562860 0.9275470 0.3062527
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5087430 0.4912570
#>  [2,] 0.5041764 0.4958236
#>  [3,] 0.4899752 0.5100248
#>  [4,] 0.4960008 0.5039992
#>  [5,] 0.4900813 0.5099187
#>  [6,] 0.5003016 0.4996984
#>  [7,] 0.4922228 0.5077772
#>  [8,] 0.5065123 0.4934877
#>  [9,] 0.5092172 0.4907828
#> [10,] 0.5134696 0.4865304
#> [11,] 0.5151160 0.4848840
#> [12,] 0.5113829 0.4886171
#> [13,] 0.4921462 0.5078538
#> [14,] 0.4920211 0.5079789
#> [15,] 0.5030204 0.4969796
#> [16,] 0.5046757 0.4953243
#> [17,] 0.4891944 0.5108056
#> [18,] 0.5022957 0.4977043
#> [19,] 0.5011720 0.4988280
#> [20,] 0.5088291 0.4911709
#> [21,] 0.5048122 0.4951878
#> [22,] 0.4930486 0.5069514
#> [23,] 0.5004819 0.4995181
#> [24,] 0.5061300 0.4938700
#> [25,] 0.4952754 0.5047246
#> [26,] 0.4896161 0.5103839
#> [27,] 0.4996095 0.5003905
#> [28,] 0.5011331 0.4988669
#> [29,] 0.5002796 0.4997204
#> [30,] 0.4965422 0.5034578
#> [31,] 0.4990424 0.5009576
#> [32,] 0.4902950 0.5097050
#> [33,] 0.5005731 0.4994269
#> [34,] 0.5067817 0.4932183
#> [35,] 0.5030654 0.4969346
#> [36,] 0.5045488 0.4954512
#> [37,] 0.4890332 0.5109668
#> [38,] 0.4894732 0.5105268
#> [39,] 0.4898632 0.5101368
#> [40,] 0.4875178 0.5124822
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]      [,3]        [,4]         [,5]         [,6]
#>  [1,]  0.04612656 1.0758095 1.2118508  0.07188930 -0.105674170 -0.379959792
#>  [2,]  0.99304909 1.3175277 1.6816452  0.15642215  0.058239346 -0.004954230
#>  [3,]  1.01624576 1.5765714 1.6422182  0.60378690  0.533548245  0.192619195
#>  [4,]  0.52815777 1.2895889 1.3346560  0.36223137  0.270601522 -0.077468112
#>  [5,]  0.99446374 1.6414280 1.7871840  0.68517914  0.563258858  0.172523419
#>  [6,]  0.66328469 1.1801232 1.2854779  0.21284197  0.014746367 -0.102973227
#>  [7,]  0.89325387 1.5442561 1.7170167  0.65837424  0.408534252  0.167786840
#>  [8,]  0.89048088 1.2169147 1.5466077  0.10776262 -0.109508792 -0.058022644
#>  [9,] -0.36055597 0.4322467 0.3620905 -0.50376733 -0.699779002 -0.593182694
#> [10,] -0.06888189 0.8044469 0.6979601 -0.34842813 -0.671496306 -0.472272419
#> [11,]  0.62956741 0.9175485 1.3048320 -0.20612895 -0.329040407 -0.262456326
#> [12,] -0.01904991 0.7705247 0.5837406 -0.37208415 -0.635732670 -0.442034143
#> [13,]  1.14715076 1.8449657 1.8121691  0.65316296  0.424363401  0.222044175
#> [14,]  1.06480576 1.7567941 1.7122818  0.56749243  0.390009795  0.196989190
#> [15,]  0.94494151 1.5052547 1.7136614  0.36878676  0.227290719 -0.013653021
#> [16,]  0.71512066 1.4782905 1.6564548  0.20703077  0.181707962 -0.224603649
#> [17,]  1.06711791 1.6571846 1.7905393  0.71063750  0.583878693  0.225508931
#> [18,]  0.75367769 1.3286116 1.5844846  0.27687882  0.189057780 -0.051154128
#> [19,]  0.57042231 1.1638300 1.1718919  0.21300587 -0.042564048 -0.028791347
#> [20,]  0.59037773 0.6363571 0.9502582 -0.13070060 -0.430424041 -0.190404553
#> [21,]  0.89560515 1.3009629 1.6989654  0.25159119 -0.005835867 -0.044084199
#> [22,]  0.82046630 1.5078949 1.5409153  0.51634855  0.446913856  0.060655467
#> [23,]  1.00264773 1.3232483 1.7391341  0.45395935  0.286847495  0.084970839
#> [24,]  0.17353169 0.7225887 0.8678541 -0.20353339 -0.475661764 -0.268401320
#> [25,]  0.21329462 0.8054705 0.7973439  0.11824664 -0.174758440 -0.144076802
#> [26,]  1.09421750 1.6218311 1.7466003  0.77082524  0.392770793  0.292288532
#> [27,]  1.05378757 1.5786992 1.8732781  0.43643612  0.366020115  0.105040786
#> [28,]  0.97918534 1.1890976 1.5251676  0.26182046  0.115169177  0.027215563
#> [29,]  0.90109897 1.2625207 1.5748583  0.24161788  0.122380121  0.024836879
#> [30,]  0.92334869 1.3686485 1.5122283  0.31751645  0.139544396  0.088312109
#> [31,]  0.36835767 1.3503310 1.3350614  0.38593929  0.115548505 -0.116193499
#> [32,]  1.02570941 1.3745918 1.5522347  0.59047228  0.329978971  0.256985376
#> [33,]  0.94042911 1.5334856 1.7571713  0.43756811  0.266546330 -0.021587426
#> [34,]  0.81457486 0.8130874 1.2796256 -0.01267598 -0.294375999 -0.082580044
#> [35,]  0.25621248 0.8538778 1.0923206 -0.04205251 -0.225633478 -0.284188595
#> [36,]  0.87109068 1.2447687 1.5828938  0.20676996 -0.003861319 -0.005624213
#> [37,]  0.79715501 1.5718891 1.6453604  0.78735399  0.497663208  0.212956703
#> [38,]  0.79220837 1.4436529 1.5058151  0.57398924  0.459413763  0.119052616
#> [39,]  1.05440559 1.7556258 1.8818134  0.84716704  0.731561508  0.226793066
#> [40,]  1.25990766 1.9169007 1.9980672  0.90034495  0.697326854  0.362205639
#>            [,7]        [,8]     [,9]      [,10]
#>  [1,] 1.4150281 -0.38239531 1.674272 -1.6310114
#>  [2,] 1.7473634 -0.52807865 2.189479 -0.7714027
#>  [3,] 1.7840312  0.11227521 1.972839 -0.6563156
#>  [4,] 1.5005314 -0.00474312 1.713826 -0.9631565
#>  [5,] 1.9544690  0.14666262 2.170072 -0.8231867
#>  [6,] 1.3371422 -0.27771494 1.758220 -0.8518513
#>  [7,] 2.0814496  0.12978295 2.187151 -0.9863828
#>  [8,] 1.6160476 -0.60757735 2.105110 -0.9068555
#>  [9,] 0.6818664 -0.27093889 0.906275 -1.4981568
#> [10,] 0.9244939 -0.56739106 1.383398 -1.5053729
#> [11,] 1.2583289 -1.08363304 1.820995 -0.9086580
#> [12,] 0.8222875 -0.54782125 1.170224 -1.2522153
#> [13,] 2.3487687  0.08778949 2.514163 -1.0120294
#> [14,] 2.1854823  0.12281622 2.352571 -0.9614668
#> [15,] 1.8665934 -0.48909432 2.319765 -1.0580790
#> [16,] 1.7778518 -0.49652151 2.088268 -1.0775044
#> [17,] 1.9866080  0.17692145 2.221547 -0.8181810
#> [18,] 1.8279758 -0.35195198 2.080397 -1.0276542
#> [19,] 1.4765837 -0.20437283 1.810227 -1.1471596
#> [20,] 0.9143636 -0.75964405 1.557560 -0.7035246
#> [21,] 1.7850658 -0.51819616 2.307381 -0.9135703
#> [22,] 1.7402434  0.03676366 1.958069 -0.8652459
#> [23,] 1.8726519 -0.39423384 2.211673 -0.8998621
#> [24,] 1.2015155 -0.20279553 1.514510 -1.3402342
#> [25,] 1.1603286  0.21752948 1.244844 -1.0528311
#> [26,] 2.1566324  0.23850521 2.425620 -1.0937833
#> [27,] 2.0494423 -0.32423498 2.377923 -0.9175476
#> [28,] 1.7360169 -0.40795959 2.059448 -0.6787628
#> [29,] 1.6335048 -0.33806465 1.955511 -0.5932291
#> [30,] 1.6405258 -0.11040030 1.975766 -0.7520713
#> [31,] 1.7136049  0.01799109 1.951247 -1.5023319
#> [32,] 1.7485894  0.16738548 2.034779 -0.8174871
#> [33,] 1.8712230 -0.36509892 2.296369 -1.0195258
#> [34,] 1.2762223 -0.65843938 1.774627 -0.6201610
#> [35,] 1.0806796 -0.19168683 1.446110 -1.1750639
#> [36,] 1.6859771 -0.49180321 2.133753 -0.9786577
#> [37,] 1.8976525  0.29122080 2.095562 -0.9534807
#> [38,] 1.6369075  0.21939637 1.793405 -0.7341591
#> [39,] 2.1497384  0.11568358 2.310791 -0.9832029
#> [40,] 2.3936863  0.22610253 2.549605 -0.8176974
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1404775 0.10550209
#>  [2,] 0.1355941 0.11888829
#>  [3,] 0.1264346 0.16653905
#>  [4,] 0.1277900 0.14378719
#>  [5,] 0.1292018 0.16888162
#>  [6,] 0.1268915 0.12568505
#>  [7,] 0.1284093 0.15952042
#>  [8,] 0.1350568 0.10900609
#>  [9,] 0.1081937 0.07132081
#> [10,] 0.1258431 0.07195181
#> [11,] 0.1401761 0.07969384
#> [12,] 0.1217213 0.07618172
#> [13,] 0.1303740 0.16179176
#> [14,] 0.1273457 0.15926397
#> [15,] 0.1433154 0.13123342
#> [16,] 0.1442801 0.12557684
#> [17,] 0.1277978 0.17102705
#> [18,] 0.1371205 0.12793754
#> [19,] 0.1257364 0.12104848
#> [20,] 0.1197696 0.08444969
#> [21,] 0.1368824 0.11763312
#> [22,] 0.1293405 0.15714793
#> [23,] 0.1379494 0.13602167
#> [24,] 0.1128416 0.08832042
#> [25,] 0.1041401 0.12303898
#> [26,] 0.1240164 0.16555803
#> [27,] 0.1405435 0.14210540
#> [28,] 0.1291946 0.12466205
#> [29,] 0.1296212 0.12850268
#> [30,] 0.1245352 0.13836651
#> [31,] 0.1311401 0.13497069
#> [32,] 0.1191872 0.15801192
#> [33,] 0.1406927 0.13840032
#> [34,] 0.1228057 0.09567742
#> [35,] 0.1206254 0.10836362
#> [36,] 0.1352811 0.11708557
#> [37,] 0.1262681 0.17014250
#> [38,] 0.1226615 0.16477508
#> [39,] 0.1356678 0.17622055
#> [40,] 0.1306062 0.18054544
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4912570 -0.4912570
#>  [2,]  0.4958236 -0.4958236
#>  [3,] -0.4899752  0.4899752
#>  [4,]  0.5039992 -0.5039992
#>  [5,]  0.5099187 -0.5099187
#>  [6,] -0.5003016  0.5003016
#>  [7,] -0.4922228  0.4922228
#>  [8,]  0.4934877 -0.4934877
#>  [9,] -0.5092172  0.5092172
#> [10,]  0.4865304 -0.4865304
#> [11,] -0.5151160  0.5151160
#> [12,]  0.4886171 -0.4886171
#> [13,]  0.5078538 -0.5078538
#> [14,]  0.5079789 -0.5079789
#> [15,]  0.4969796 -0.4969796
#> [16,]  0.4953243 -0.4953243
#> [17,] -0.4891944  0.4891944
#> [18,] -0.5022957  0.5022957
#> [19,] -0.5011720  0.5011720
#> [20,]  0.4911709 -0.4911709
#> [21,] -0.5048122  0.5048122
#> [22,]  0.5069514 -0.5069514
#> [23,]  0.4995181 -0.4995181
#> [24,] -0.5061300  0.5061300
#> [25,] -0.4952754  0.4952754
#> [26,] -0.4896161  0.4896161
#> [27,]  0.5003905 -0.5003905
#> [28,] -0.5011331  0.5011331
#> [29,]  0.4997204 -0.4997204
#> [30,] -0.4965422  0.4965422
#> [31,]  0.5009576 -0.5009576
#> [32,] -0.4902950  0.4902950
#> [33,]  0.4994269 -0.4994269
#> [34,]  0.4932183 -0.4932183
#> [35,] -0.5030654  0.5030654
#> [36,]  0.4954512 -0.4954512
#> [37,] -0.4890332  0.4890332
#> [38,] -0.4894732  0.4894732
#> [39,]  0.5101368 -0.5101368
#> [40,] -0.4875178  0.4875178
#> 
#> $L
#> [1] 0.7130544 0.7718096 0.7436739 0.6941014 0.6881489 0.6896172
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3768116