Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.3124180 -0.01462231 -0.06186708  0.003380005 -0.02792232  0.003363308
#>  [2,] -0.2043090 -0.22232168 -0.18805043 -0.168128525 -0.06079223 -0.111321759
#>  [3,] -0.3921797 -0.16475409 -0.07326068 -0.205367595 -0.07832600 -0.123071803
#>  [4,] -0.2445542 -0.13094998 -0.14645741 -0.036312587 -0.03617295  0.008307047
#>  [5,] -0.2748618 -0.07306023 -0.10013723 -0.065845295 -0.07338906 -0.033482043
#>  [6,] -0.2712897 -0.08028390 -0.13766900 -0.037515963 -0.10178070 -0.024847666
#>  [7,] -0.3048454 -0.16209340 -0.08453254 -0.155879899  0.01292484 -0.046249458
#>  [8,] -0.2989969 -0.06459108 -0.15868287 -0.173239730 -0.04904727 -0.066538068
#>  [9,] -0.2417180 -0.08527728 -0.04053784 -0.136888441 -0.17258641 -0.057190784
#> [10,] -0.2327474 -0.11179228 -0.03206484 -0.039051125 -0.15140960 -0.059693105
#>                V15          V16         V17          V18          V19
#>  [1,]  0.083464092  0.057920356  0.09592434  0.051321739  0.111261510
#>  [2,] -0.095277911 -0.037628019 -0.08416082 -0.058051355 -0.096793472
#>  [3,] -0.118284819 -0.047291034 -0.01485474 -0.007689975 -0.099193063
#>  [4,]  0.066049386  0.131599498  0.14898478  0.184364734  0.198488418
#>  [5,] -0.039521464 -0.068844571  0.07311722 -0.056299412  0.006663702
#>  [6,] -0.168832114 -0.041228422  0.02738419  0.030615305 -0.044322550
#>  [7,] -0.002758411 -0.034416484  0.06222088  0.062961318  0.004506121
#>  [8,] -0.126739257 -0.131661051 -0.08152800 -0.058761543 -0.052417926
#>  [9,] -0.118188325 -0.001127863 -0.02428308 -0.019529569  0.053474818
#> [10,]  0.035686576 -0.057498752 -0.04557751  0.124145025  0.068886038
#>               V2          V20         V21        V22           V23          V24
#>  [1,] -0.3711392  0.118068517  0.18419565 0.07810898  0.1452852133  0.292260677
#>  [2,] -0.2484579  0.087546927  0.10825109 0.04511105  0.0684413250  0.101632306
#>  [3,] -0.3674687 -0.097517292  0.05944827 0.05806141  0.0848442869  0.008294162
#>  [4,] -0.1632758  0.154842138  0.08959646 0.19503770  0.1683415096  0.069583433
#>  [5,] -0.2273188 -0.008595002  0.01071725 0.07066772 -0.0001730217  0.159758500
#>  [6,] -0.2233954  0.077339030  0.09947698 0.04675765  0.0926162286  0.173107805
#>  [7,] -0.2341821  0.131564714  0.16694223 0.12996430  0.1122192051  0.197332407
#>  [8,] -0.3083293  0.028905441 -0.04061372 0.14303462  0.0777070110 -0.005376307
#>  [9,] -0.2081185  0.045962888  0.06532154 0.04164073  0.0783957411  0.086998866
#> [10,] -0.3718017  0.058280503  0.20271162 0.15949378  0.1692240423  0.144794064
#>              V25         V26        V27       V28          V29         V3
#>  [1,] 0.20850438 0.260164029 0.18058675 0.1323744  0.252438638 -0.2746363
#>  [2,] 0.04807277 0.098458876 0.13347755 0.2595292  0.110266651 -0.2147479
#>  [3,] 0.19690912 0.248555222 0.28415915 0.2546021  0.132631685 -0.3869166
#>  [4,] 0.13137093 0.137084164 0.03840253 0.1012616  0.005698975 -0.2402306
#>  [5,] 0.16261202 0.043683952 0.14300306 0.1115883  0.008882057 -0.2315791
#>  [6,] 0.19910188 0.163456473 0.11923841 0.0745537  0.086880095 -0.2930387
#>  [7,] 0.21994573 0.110688963 0.22863161 0.2116232  0.119839371 -0.2753572
#>  [8,] 0.08873593 0.182917839 0.21543256 0.1956014  0.109604853 -0.2055000
#>  [9,] 0.12720648 0.002162595 0.01604970 0.1044501 -0.030261014 -0.3651459
#> [10,] 0.12662079 0.198637887 0.25512269 0.2138443  0.278780705 -0.3426422
#>                V30          V31          V32          V33          V34
#>  [1,]  0.233458876  0.126645426  0.014451199  0.077350993  0.074220173
#>  [2,]  0.119582708 -0.087613823 -0.038464183 -0.131986377 -0.080733188
#>  [3,] -0.000165498  0.096241954  0.001637086  0.020444035 -0.008133419
#>  [4,] -0.044250760  0.060318798  0.090265725  0.002233399 -0.049904559
#>  [5,]  0.122098058  0.101486732  0.002385554 -0.074000378  0.035611839
#>  [6,] -0.073117881 -0.047842842 -0.065954910 -0.130219962 -0.196510769
#>  [7,]  0.114482498  0.097007317  0.092614584  0.089386274  0.058650757
#>  [8,]  0.073458735 -0.023772543  0.038409811  0.030645279 -0.095322222
#>  [9,]  0.028308864  0.063977517 -0.143069213 -0.022667075 -0.128796290
#> [10,]  0.038343675  0.003924863  0.047486589 -0.015761430  0.085675903
#>                V35          V36          V37          V38         V39
#>  [1,]  0.021261850  0.027560433  0.013147288 -0.003433092 -0.01828644
#>  [2,]  0.007937041 -0.005530206 -0.052197694  0.068907045 -0.01889105
#>  [3,] -0.034377690  0.051923230 -0.103060262 -0.040379801 -0.13217206
#>  [4,] -0.012609342 -0.061133532 -0.008487962 -0.106245134 -0.12518373
#>  [5,] -0.086852537 -0.096077033  0.017073516  0.084591912  0.04179336
#>  [6,] -0.225299524 -0.150230141 -0.136417910 -0.089882635 -0.14702783
#>  [7,]  0.027190731 -0.137014489 -0.178010592 -0.165654923 -0.08545293
#>  [8,]  0.063273486 -0.024880497 -0.051564081  0.020388449  0.05666754
#>  [9,] -0.120521502  0.055504084  0.022562790 -0.186652703 -0.06257444
#> [10,]  0.117776668  0.098896429 -0.013782434 -0.097901371  0.06590914
#>               V4         V40         V41         V42         V43           V44
#>  [1,] -0.3316516  0.08145259 -0.02252749  0.05690760 -0.04812212 -0.0855551862
#>  [2,] -0.1671708 -0.10162493 -0.19627993 -0.07886305 -0.17717586 -0.1678638078
#>  [3,] -0.2154896 -0.17495834 -0.13837314 -0.19164743 -0.13114069 -0.0388100469
#>  [4,] -0.2770800  0.01958048 -0.13233242 -0.05726082 -0.05262055 -0.1175188423
#>  [5,] -0.3251246 -0.05208183  0.01910177  0.01643678 -0.11465577 -0.1583162717
#>  [6,] -0.3258932 -0.04538701 -0.14953405 -0.04210018 -0.19425189 -0.0850092999
#>  [7,] -0.3531458 -0.12360848  0.02622109  0.02443802 -0.02806046 -0.1248738754
#>  [8,] -0.2099848 -0.11894632 -0.08862671 -0.07631724  0.05048850 -0.0002912991
#>  [9,] -0.1909151 -0.10828556 -0.13508115 -0.10311699 -0.12775505 -0.2030082856
#> [10,] -0.1909809  0.01941961 -0.04812380 -0.09232951 -0.12219702 -0.0924299849
#>               V45         V46         V47        V48        V49         V5
#>  [1,] -0.17855577 -0.05947825 -0.21742877 -0.1609346 -0.2989137 -0.2369651
#>  [2,] -0.20012001 -0.11302640 -0.24025627 -0.2521851 -0.3495562 -0.2278322
#>  [3,] -0.10398035 -0.16817952 -0.27685460 -0.2394660 -0.2888086 -0.2811430
#>  [4,] -0.18403258 -0.10102271 -0.06906325 -0.1186909 -0.1597776 -0.2290341
#>  [5,] -0.03075891 -0.13290951 -0.13108103 -0.1879016 -0.1721140 -0.3037906
#>  [6,] -0.22681010 -0.14809295 -0.09058695 -0.2305079 -0.1683192 -0.3254457
#>  [7,] -0.17351214 -0.21039076 -0.18770652 -0.2120801 -0.3000973 -0.1450355
#>  [8,] -0.07520465 -0.18320079 -0.11419934 -0.1677616 -0.3071991 -0.2700317
#>  [9,] -0.10992677 -0.26789724 -0.23007715 -0.1755980 -0.2570943 -0.1625024
#> [10,] -0.07042698 -0.13798320 -0.24376199 -0.2967783 -0.2255113 -0.2004383
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3668777 -0.2804324 -0.4062191 -0.4019477 -0.4015169 -0.4032668
#>  [2,] -0.2765210 -0.2352184 -0.2063630 -0.2838948 -0.3143822 -0.3054729
#>  [3,] -0.2863025 -0.3339224 -0.3556997 -0.2892743 -0.3412191 -0.4227266
#>  [4,] -0.3721307 -0.2838891 -0.2440940 -0.3027707 -0.3092226 -0.2503997
#>  [5,] -0.3910076 -0.4123781 -0.4034669 -0.3478871 -0.3618413 -0.3488314
#>  [6,] -0.1909536 -0.2049852 -0.2133203 -0.2458615 -0.3601552 -0.3288795
#>  [7,] -0.3114631 -0.3979706 -0.3039939 -0.3497358 -0.4133125 -0.3024991
#>  [8,] -0.3065439 -0.2421949 -0.3964558 -0.3511753 -0.4134063 -0.3280707
#>  [9,] -0.4002956 -0.2288459 -0.3238261 -0.4262541 -0.2318084 -0.2979306
#> [10,] -0.2252853 -0.3728538 -0.3423102 -0.3200715 -0.2429767 -0.3032629
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3567728 -0.3701601 -0.2875567 -0.3621186 -0.17535304 -0.3859042
#>  [2,] -0.3263531 -0.2184281 -0.3916616 -0.2408198 -0.25677969 -0.2643501
#>  [3,] -0.2748695 -0.4456429 -0.3151524 -0.3081583 -0.17275488 -0.3762672
#>  [4,] -0.3967705 -0.2665454 -0.4077990 -0.4000341 -0.16238335 -0.3089151
#>  [5,] -0.3873132 -0.2639859 -0.2824175 -0.2981285 -0.18260121 -0.3523066
#>  [6,] -0.2686447 -0.3848023 -0.2509099 -0.2940000 -0.09698254 -0.2878147
#>  [7,] -0.3002012 -0.3051236 -0.4395564 -0.3802280 -0.11504412 -0.4071965
#>  [8,] -0.3198746 -0.2799052 -0.3884427 -0.3961495 -0.28923606 -0.3908305
#>  [9,] -0.2904417 -0.3711176 -0.2606495 -0.2497130 -0.31534342 -0.4099523
#> [10,] -0.4223777 -0.3365480 -0.3657169 -0.4270798 -0.17348650 -0.4108010
#>               V7          V8          V9
#>  [1,] -0.1849952 -0.15458160 -0.07411080
#>  [2,] -0.1994854 -0.09879943 -0.12398261
#>  [3,] -0.2131105 -0.18408400 -0.26669607
#>  [4,] -0.1503024 -0.24941341 -0.07886080
#>  [5,] -0.1671220 -0.24430225 -0.05310056
#>  [6,] -0.2162343 -0.12656408 -0.14316618
#>  [7,] -0.2032950 -0.16171264 -0.16295592
#>  [8,] -0.2202861 -0.14089891 -0.19779502
#>  [9,] -0.1288224 -0.12575044 -0.18973316
#> [10,] -0.2389213 -0.13005216 -0.13290245
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] -0.07653906 -0.08827755 -0.13236093 0.08614581 -0.09640256  0.05003204
#> [2,]  0.04128950  0.05694322 -0.01573428 0.02673844  0.04689681 -0.01397631
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] 0.072117906 0.005696489 -0.07715989  0.09203862
#> [2,] 0.005642109 0.048598258  0.03588572 -0.06677775
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  6.833829e-05  3.901081e-04  5.143647e-04  4.935823e-04  4.817280e-04
#>  [2,]  1.097264e-04  8.361423e-04  1.237153e-03  1.217701e-03  1.082677e-03
#>  [3,]  9.570600e-05  6.576812e-04  8.870021e-04  8.756613e-04  8.485731e-04
#>  [4,]  4.662412e-05  1.602692e-04 -7.644207e-05  7.875397e-06  1.852650e-04
#>  [5,]  1.284922e-04  9.050293e-04  1.280174e-03  1.246013e-03  1.152732e-03
#>  [6,]  2.723726e-05  8.266363e-05 -2.184161e-04 -2.016630e-04  1.507005e-05
#>  [7,]  4.564296e-05  1.507694e-04 -7.880302e-05 -6.030864e-06  1.578569e-04
#>  [8,]  7.314087e-05  4.288291e-04  4.355661e-04  4.488821e-04  5.201002e-04
#>  [9,]  8.530350e-05  6.421653e-04  9.345642e-04  9.324270e-04  8.405593e-04
#> [10,] -1.446253e-05 -2.053494e-04 -6.138289e-04 -5.160834e-04 -3.026523e-04
#>                V14          V15          V16          V17          V18
#>  [1,] 4.674409e-04 0.0004446484 0.0004310812 0.0004526239 0.0005098042
#>  [2,] 8.063590e-04 0.0005309330 0.0003664203 0.0003107417 0.0005418724
#>  [3,] 7.746072e-04 0.0006865884 0.0006635785 0.0007067289 0.0008312005
#>  [4,] 5.218942e-04 0.0008815781 0.0011446011 0.0014086392 0.0014058405
#>  [5,] 9.502633e-04 0.0007227560 0.0005983897 0.0005787467 0.0007913183
#>  [6,] 4.113158e-04 0.0007715956 0.0010541069 0.0013347716 0.0012527672
#>  [7,] 4.688231e-04 0.0008128545 0.0010927271 0.0013914115 0.0013615791
#>  [8,] 6.472748e-04 0.0007756403 0.0008946031 0.0010532358 0.0010979383
#>  [9,] 6.567776e-04 0.0004880395 0.0003926897 0.0003782887 0.0005495604
#> [10,] 9.850232e-05 0.0005690169 0.0009049119 0.0012247877 0.0011619902
#>                V19            V2           V20          V21          V22
#>  [1,] 0.0007230687  9.551238e-05  0.0008686190 0.0008559744 0.0007630467
#>  [2,] 0.0012892678  1.719312e-04  0.0017681811 0.0016281793 0.0015769619
#>  [3,] 0.0012495441  1.393532e-04  0.0015259451 0.0015436486 0.0015076982
#>  [4,] 0.0010209820  4.035905e-05  0.0006689061 0.0010386360 0.0008837357
#>  [5,] 0.0014993426  1.930314e-04  0.0019698501 0.0018598448 0.0017586029
#>  [6,] 0.0007683523  1.264054e-05  0.0004540012 0.0008557138 0.0007693626
#>  [7,] 0.0009460631  3.648654e-05  0.0005427020 0.0009384949 0.0008191263
#>  [8,] 0.0011223689  9.161694e-05  0.0011021322 0.0013006581 0.0012170609
#>  [9,] 0.0010717018  1.321375e-04  0.0013954621 0.0013379055 0.0013041716
#> [10,] 0.0004389830 -5.334777e-05 -0.0002241050 0.0002913822 0.0001975115
#>                V23          V24          V25          V26          V27
#>  [1,] 7.157976e-04 0.0006317470 0.0005925687 0.0006437887 0.0006694527
#>  [2,] 1.521235e-03 0.0010870225 0.0007471033 0.0008337967 0.0008753397
#>  [3,] 1.455691e-03 0.0012721675 0.0011473072 0.0012413314 0.0012386001
#>  [4,] 7.989480e-04 0.0013068196 0.0017231906 0.0018443161 0.0019504380
#>  [5,] 1.694812e-03 0.0013162646 0.0010287446 0.0011372942 0.0011778183
#>  [6,] 7.415185e-04 0.0013142990 0.0017982660 0.0018952514 0.0019414730
#>  [7,] 7.129161e-04 0.0012414599 0.0016904000 0.0017736177 0.0018676666
#>  [8,] 1.144891e-03 0.0013070025 0.0014666278 0.0015557575 0.0015852207
#>  [9,] 1.245247e-03 0.0009652430 0.0007482140 0.0008186736 0.0008383270
#> [10,] 7.114862e-05 0.0008402253 0.0014896991 0.0015316792 0.0015900965
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0007015204 0.0006358182  1.172998e-04 0.0004907665 3.403115e-04
#>  [2,] 0.0009157254 0.0008064310  1.900700e-04 0.0004040735 5.162456e-07
#>  [3,] 0.0012212651 0.0011263217  1.622592e-04 0.0008304222 5.400091e-04
#>  [4,] 0.0020417039 0.0020776525  8.133526e-05 0.0019618687 1.800704e-03
#>  [5,] 0.0012269521 0.0011027237  2.189966e-04 0.0006866357 2.694679e-04
#>  [6,] 0.0019362128 0.0018417735  5.223068e-05 0.0017508680 1.684189e-03
#>  [7,] 0.0019351281 0.0019304246  7.969951e-05 0.0018874817 1.778212e-03
#>  [8,] 0.0015799633 0.0014842991  1.259702e-04 0.0012918414 1.090294e-03
#>  [9,] 0.0008364027 0.0007426941  1.482511e-04 0.0004333923 1.265252e-04
#> [10,] 0.0016129166 0.0016379109 -2.141938e-05 0.0016962025 1.674602e-03
#>                V32           V33           V34           V35           V36
#>  [1,] 0.0003128640  2.148135e-04  0.0001413080  1.937082e-05 -0.0001603428
#>  [2,] 0.0002226277 -8.244208e-05 -0.0003757958 -9.229521e-04 -0.0014204557
#>  [3,] 0.0005240643  3.411815e-04  0.0002033596 -4.968458e-05 -0.0003046784
#>  [4,] 0.0010601511  1.201692e-03  0.0014391276  1.964905e-03  0.0024195131
#>  [5,] 0.0004118640  1.389855e-04 -0.0001285853 -6.287703e-04 -0.0011269025
#>  [6,] 0.0010233314  1.193565e-03  0.0013673779  1.893648e-03  0.0023141650
#>  [7,] 0.0010270238  1.125683e-03  0.0013477384  1.927217e-03  0.0024292733
#>  [8,] 0.0007464371  7.024531e-04  0.0007217802  8.529122e-04  0.0009297940
#>  [9,] 0.0002280191  3.272358e-06 -0.0001920696 -5.452824e-04 -0.0008634189
#> [10,] 0.0007559740  1.001992e-03  0.0013842058  2.270463e-03  0.0031214384
#>                 V37           V38          V39            V4          V40
#>  [1,] -6.932929e-05  1.545156e-04 0.0002805789  1.629258e-04 1.784641e-04
#>  [2,] -1.126609e-03 -2.742791e-04 0.0002693690  2.984540e-04 9.180618e-05
#>  [3,] -1.600468e-04  2.414154e-04 0.0004936983  2.297030e-04 3.775248e-04
#>  [4,]  2.358527e-03  1.727279e-03 0.0012964599  3.073571e-05 1.180881e-03
#>  [5,] -8.602257e-04 -7.353684e-05 0.0004201776  3.302827e-04 2.097931e-04
#>  [6,]  2.248378e-03  1.481127e-03 0.0009666827  3.220665e-06 9.227375e-04
#>  [7,]  2.409550e-03  1.742307e-03 0.0012781584  1.832829e-05 1.222321e-03
#>  [8,]  1.010717e-03  8.974304e-04 0.0007995030  1.385476e-04 7.110129e-04
#>  [9,] -6.444239e-04 -7.751694e-05 0.0002884465  2.215490e-04 1.760676e-04
#> [10,]  3.031729e-03  1.981577e-03 0.0012588138 -1.397915e-04 1.285656e-03
#>                V41          V42          V43          V44           V45
#>  [1,] 0.0002178726 0.0003260173 0.0003281699 3.175022e-04  4.028429e-04
#>  [2,] 0.0002113502 0.0005422113 0.0006261169 7.008411e-04  9.814541e-04
#>  [3,] 0.0004238551 0.0006058867 0.0006022053 5.863059e-04  7.162314e-04
#>  [4,] 0.0009434068 0.0007174569 0.0005142770 2.295158e-04 -1.355441e-04
#>  [5,] 0.0003202346 0.0006474212 0.0007057637 7.471402e-04  1.020236e-03
#>  [6,] 0.0007369773 0.0004913845 0.0002839515 6.174963e-05 -2.174121e-04
#>  [7,] 0.0010189728 0.0007619974 0.0005250145 2.578909e-04 -9.485072e-05
#>  [8,] 0.0006482200 0.0006337631 0.0005320203 4.142573e-04  3.352036e-04
#>  [9,] 0.0002425879 0.0004701944 0.0005124811 5.447942e-04  7.189769e-04
#> [10,] 0.0009819038 0.0005809939 0.0003141927 1.208888e-06 -5.494065e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0003670267  2.694722e-04  1.910162e-04  1.063077e-04  1.694998e-04
#>  [2,]  0.0009046717  6.326242e-04  4.365827e-04  2.650805e-04  3.301620e-04
#>  [3,]  0.0006498299  4.698107e-04  3.306357e-04  1.937703e-04  2.704332e-04
#>  [4,] -0.0001647853 -9.166525e-06  9.691539e-06 -2.013274e-05  9.389231e-05
#>  [5,]  0.0009370164  6.569177e-04  4.579691e-04  2.731605e-04  3.612198e-04
#>  [6,] -0.0002254454 -7.941664e-05 -3.746779e-05 -4.875904e-05  4.286424e-05
#>  [7,] -0.0001384875  1.352201e-05  2.547601e-05 -1.289536e-05  8.477009e-05
#>  [8,]  0.0002826717  2.545769e-04  1.870946e-04  9.590443e-05  1.847596e-04
#>  [9,]  0.0006653079  4.809429e-04  3.317223e-04  1.992911e-04  2.533685e-04
#> [10,] -0.0005728656 -2.702023e-04 -1.747848e-04 -1.318381e-04 -6.247144e-05
#>                 V50           V51           V52          V53          V54
#>  [1,]  3.032173e-05  3.489197e-05  3.251527e-05 2.093714e-05 1.789431e-05
#>  [2,]  5.654899e-05  7.674149e-05  5.884264e-05 2.679823e-05 3.006914e-05
#>  [3,]  5.239573e-05  5.803358e-05  4.766518e-05 2.970297e-05 2.878584e-05
#>  [4,]  2.346726e-05 -8.191760e-07  1.107399e-05 3.875674e-05 2.370152e-05
#>  [5,]  6.530490e-05  8.219554e-05  6.662933e-05 3.539488e-05 3.602478e-05
#>  [6,]  2.083232e-05 -8.791762e-06  6.087605e-06 3.006678e-05 1.688598e-05
#>  [7,]  2.390141e-05 -3.498387e-06  9.328922e-06 3.816789e-05 2.227435e-05
#>  [8,]  4.001942e-05  2.986726e-05  3.087311e-05 3.335516e-05 2.516395e-05
#>  [9,]  4.465535e-05  5.740182e-05  4.398361e-05 2.241362e-05 2.408369e-05
#> [10,] -6.709291e-06 -4.198433e-05 -2.294849e-05 2.427487e-05 6.456053e-06
#>                V55          V56           V57          V58          V59
#>  [1,] 1.569279e-05 1.142268e-05  8.310199e-06 1.421733e-05 9.796283e-06
#>  [2,] 2.240628e-05 1.735441e-05 -1.126869e-06 2.431468e-05 5.989651e-06
#>  [3,] 2.416385e-05 1.808354e-05  1.008444e-05 2.281773e-05 1.395506e-05
#>  [4,] 2.456843e-05 1.929666e-05  3.379703e-05 1.903913e-05 3.610173e-05
#>  [5,] 2.886484e-05 2.106365e-05  5.294594e-06 2.847156e-05 1.164286e-05
#>  [6,] 2.012854e-05 1.413692e-05  3.353624e-05 9.965417e-06 2.838423e-05
#>  [7,] 2.306248e-05 1.890857e-05  3.427274e-05 1.858228e-05 3.493551e-05
#>  [8,] 2.357216e-05 1.824938e-05  2.186681e-05 1.937290e-05 2.265924e-05
#>  [9,] 1.828670e-05 1.433648e-05  1.646170e-06 1.939286e-05 7.261600e-06
#> [10,] 1.012737e-05 1.018166e-05  3.156320e-05 5.524303e-06 3.152317e-05
#>                 V6          V60           V7           V8            V9
#>  [1,] 1.490921e-04 7.716802e-06 0.0001504998 0.0001557082  3.086074e-04
#>  [2,] 2.837077e-04 7.253462e-06 0.0002551532 0.0002384377  6.081743e-04
#>  [3,] 2.650826e-04 1.059945e-05 0.0002698407 0.0002710663  5.066764e-04
#>  [4,] 1.932706e-04 2.024417e-05 0.0002589180 0.0003279686  2.416560e-04
#>  [5,] 3.163312e-04 1.065748e-05 0.0002989918 0.0002926157  6.761564e-04
#>  [6,] 1.336643e-04 1.417379e-05 0.0002172364 0.0002813288  1.692457e-04
#>  [7,] 1.819920e-04 1.974031e-05 0.0002521574 0.0003316125  2.395169e-04
#>  [8,] 2.199364e-04 1.384279e-05 0.0002637053 0.0002981087  3.832979e-04
#>  [9,] 2.301514e-04 7.079757e-06 0.0002206491 0.0002096197  4.692044e-04
#> [10,] 6.682215e-05 1.547225e-05 0.0001582087 0.0002383049 -2.159113e-05
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]      [,4]        [,5]        [,6]
#> [1,]  0.0423699  0.03502703  0.04086036  0.032803  0.03699181  0.02126978
#> [2,] -0.0423699 -0.03502703 -0.04086036 -0.032803 -0.03699181 -0.02126978
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.03484144  0.03857828  0.03142131  0.04205742
#> [2,] -0.03484144 -0.03857828 -0.03142131 -0.04205742
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 1.0119096 0.3217519 0.5065885 0.4980446 0.5773234 0.2691586 0.5218501
#>  [8] 0.6386260 0.3828389 0.7138895
#> 
#> $B[[2]]
#> [1] -0.058790086 -0.002563214
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0009378458 0.0009372178 0.0015908391 0.0029937506 0.0014338744
#>  [6] 0.0027223081 0.0028586136 0.0021922544 0.0009596324 0.0024846530
#> 
#> $vB[[2]]
#> [1]  0.04526256 -0.04526256
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#>  [2,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [3,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#>  [4,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#>  [5,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#>  [6,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#>  [7,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#>  [8,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#>  [9,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [10,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [11,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [12,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [13,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [14,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [15,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [16,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [17,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [18,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [19,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [20,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [21,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [22,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [23,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [24,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [25,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [26,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [27,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [28,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [29,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [30,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [31,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [32,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [33,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [34,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [35,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [36,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [37,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [38,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [39,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [40,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#>  [2,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [3,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#>  [4,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#>  [5,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#>  [6,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#>  [7,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#>  [8,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#>  [9,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [10,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [11,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [12,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [13,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [14,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [15,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [16,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [17,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [18,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [19,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [20,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [21,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [22,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [23,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [24,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [25,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [26,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [27,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [28,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [29,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [30,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [31,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [32,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [33,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [34,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [35,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [36,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [37,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [38,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [39,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [40,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#>  [2,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [3,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#>  [4,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#>  [5,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#>  [6,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#>  [7,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#>  [8,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#>  [9,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [10,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [11,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [12,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [13,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [14,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [15,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [16,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [17,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [18,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [19,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [20,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [21,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [22,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [23,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [24,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [25,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [26,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [27,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [28,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [29,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [30,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [31,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [32,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [33,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [34,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [35,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [36,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [37,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [38,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [39,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [40,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#>  [2,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [3,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#>  [4,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#>  [5,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#>  [6,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#>  [7,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#>  [8,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#>  [9,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [10,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [11,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [12,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [13,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [14,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [15,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [16,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [17,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [18,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [19,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [20,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [21,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [22,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [23,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [24,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [25,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [26,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [27,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [28,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [29,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [30,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [31,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [32,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [33,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [34,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [35,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [36,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [37,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [38,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [39,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [40,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#>  [2,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [3,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#>  [4,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#>  [5,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#>  [6,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#>  [7,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#>  [8,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#>  [9,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [10,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [11,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [12,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [13,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [14,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [15,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [16,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [17,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [18,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [19,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [20,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [21,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [22,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [23,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [24,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [25,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [26,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [27,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [28,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [29,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [30,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [31,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [32,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [33,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [34,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [35,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [36,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [37,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [38,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [39,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [40,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#>  [2,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [3,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#>  [4,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#>  [5,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#>  [6,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#>  [7,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#>  [8,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#>  [9,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [10,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [11,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [12,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [13,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [14,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [15,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [16,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [17,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [18,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [19,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [20,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [21,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [22,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [23,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [24,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [25,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [26,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [27,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [28,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [29,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [30,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [31,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [32,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [33,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [34,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [35,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [36,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [37,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [38,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [39,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [40,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9235091 0.5614797 0.6250816 0.7884430 0.6774079 0.4983619 0.7872768
#>  [2,] 0.8781519 0.3742053 0.4254491 0.6239543 0.6011419 0.2528733 0.6091132
#>  [3,] 0.9186577 0.4731523 0.5257645 0.8131685 0.6246587 0.5244919 0.7607245
#>  [4,] 0.9036376 0.6180313 0.7284568 0.7152606 0.7148039 0.6163173 0.7887551
#>  [5,] 0.9194379 0.5934895 0.7208476 0.7863860 0.7029083 0.5981262 0.8076905
#>  [6,] 0.8990191 0.4279719 0.4981178 0.6721343 0.6204713 0.4091481 0.6957569
#>  [7,] 0.9516598 0.6756502 0.7788046 0.7828069 0.7515484 0.5508774 0.8553552
#>  [8,] 0.9303830 0.5897078 0.6986469 0.7247233 0.6904717 0.4673324 0.7734803
#>  [9,] 0.9196753 0.5822190 0.7077492 0.7249743 0.6920040 0.4525277 0.7916095
#> [10,] 0.9495816 0.6799634 0.7841638 0.7911198 0.7527844 0.5991913 0.8573101
#> [11,] 0.9381413 0.6104437 0.7096609 0.7854416 0.7063666 0.6259422 0.8358370
#> [12,] 0.9487795 0.5897043 0.6917311 0.8353225 0.7129991 0.6239278 0.8475185
#> [13,] 0.9180266 0.4553467 0.5155042 0.8079999 0.6227488 0.5078465 0.7613830
#> [14,] 0.9364467 0.6140464 0.7081350 0.8503893 0.7141813 0.6576094 0.8461560
#> [15,] 0.9146149 0.5588928 0.6637132 0.7159966 0.7023322 0.5059773 0.7826780
#> [16,] 0.8952101 0.5790104 0.6717159 0.7413632 0.6662704 0.5596440 0.7677400
#> [17,] 0.9294814 0.5132974 0.6154859 0.7266595 0.6718265 0.4954121 0.7839228
#> [18,] 0.9140388 0.4852455 0.6518946 0.6290346 0.6668561 0.3795856 0.7384420
#> [19,] 0.9329521 0.5984163 0.7237095 0.7271700 0.7096131 0.5024200 0.8214026
#> [20,] 0.9039325 0.4273434 0.5853638 0.6028460 0.6481592 0.3211015 0.6971172
#> [21,] 0.8694992 0.4614784 0.5337417 0.5931882 0.6404138 0.3694478 0.6676912
#> [22,] 0.9136870 0.4690310 0.5624755 0.7338394 0.6471869 0.4348019 0.7193720
#> [23,] 0.8988125 0.3999818 0.5276050 0.6373460 0.6062167 0.3269247 0.7450206
#> [24,] 0.8842015 0.4997067 0.5835752 0.5956022 0.6469859 0.3065373 0.6841798
#> [25,] 0.9129838 0.5315824 0.5919956 0.8086146 0.6537087 0.5150522 0.7919827
#> [26,] 0.9247654 0.4082796 0.5077050 0.7161479 0.6452951 0.3697658 0.7729622
#> [27,] 0.9360590 0.6274178 0.7731991 0.7585319 0.7363367 0.6169021 0.8474604
#> [28,] 0.8684796 0.5512746 0.5928444 0.6394562 0.6668367 0.4355680 0.6846602
#> [29,] 0.9339343 0.6278638 0.7212121 0.7501283 0.7220003 0.5122685 0.8188374
#> [30,] 0.8764629 0.3699226 0.4445566 0.5725002 0.5963352 0.2767208 0.6117486
#> [31,] 0.8980656 0.5419340 0.6336792 0.7268023 0.6698413 0.5082213 0.7379334
#> [32,] 0.9229966 0.4878326 0.5887500 0.7567797 0.6693011 0.4824905 0.7469211
#> [33,] 0.9385564 0.6417017 0.7531385 0.7752448 0.7420473 0.6281180 0.8389105
#> [34,] 0.8902249 0.5563219 0.6193425 0.7376157 0.6563046 0.5344248 0.7580971
#> [35,] 0.8989040 0.3880412 0.4181960 0.7025631 0.6043607 0.3392856 0.6714819
#> [36,] 0.9507360 0.6255462 0.7322937 0.8697134 0.7313919 0.6425969 0.8678970
#> [37,] 0.9400515 0.5590617 0.6989094 0.7305970 0.7132516 0.4069457 0.7810877
#> [38,] 0.8847062 0.5916783 0.6295018 0.6492289 0.6794016 0.4361159 0.7056295
#> [39,] 0.9206882 0.5960318 0.6890196 0.7036738 0.6991526 0.5356780 0.8032164
#> [40,] 0.8591450 0.4638069 0.5481850 0.6604702 0.6155914 0.4323949 0.6881272
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6834623 0.5038111 0.8577483
#>  [2,] 0.5598007 0.3221065 0.7537014
#>  [3,] 0.5706208 0.4862671 0.8247999
#>  [4,] 0.7307730 0.5656816 0.8354043
#>  [5,] 0.6994319 0.5700292 0.8549058
#>  [6,] 0.5877390 0.4366890 0.7983501
#>  [7,] 0.7861313 0.5675717 0.9156453
#>  [8,] 0.7459321 0.4878278 0.8874388
#>  [9,] 0.7252724 0.4999359 0.8569883
#> [10,] 0.7811238 0.5922641 0.9116768
#> [11,] 0.7197913 0.5479677 0.8816672
#> [12,] 0.6957531 0.5473535 0.8852902
#> [13,] 0.5442735 0.4691970 0.8145009
#> [14,] 0.6957134 0.5902395 0.8762331
#> [15,] 0.6880200 0.4810100 0.8330097
#> [16,] 0.6761045 0.5183720 0.8274448
#> [17,] 0.6838986 0.4348051 0.8558986
#> [18,] 0.6962160 0.3930591 0.8298752
#> [19,] 0.7356053 0.5266868 0.8640616
#> [20,] 0.6690132 0.3434116 0.8038548
#> [21,] 0.6407662 0.3867122 0.7545164
#> [22,] 0.6029497 0.4546275 0.8157049
#> [23,] 0.6064983 0.3232797 0.7870321
#> [24,] 0.6634385 0.4052246 0.8076650
#> [25,] 0.6299331 0.5352686 0.8356718
#> [26,] 0.6037403 0.3818948 0.8348129
#> [27,] 0.7626607 0.5644935 0.8777403
#> [28,] 0.6687372 0.4879247 0.7861105
#> [29,] 0.7484974 0.5486501 0.8859534
#> [30,] 0.5911141 0.3168307 0.7536022
#> [31,] 0.6922218 0.4949247 0.8287009
#> [32,] 0.6226743 0.4854952 0.8316891
#> [33,] 0.7562024 0.5824997 0.8789515
#> [34,] 0.6596902 0.5012771 0.8292147
#> [35,] 0.5115515 0.3676143 0.7731211
#> [36,] 0.7141326 0.5979720 0.8980919
#> [37,] 0.7246694 0.4596883 0.8823200
#> [38,] 0.6771687 0.5070048 0.8160833
#> [39,] 0.7155452 0.5115014 0.8451017
#> [40,] 0.5878913 0.4382497 0.7478124
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5885840 0.4114160
#>  [2,] 0.5731167 0.4268833
#>  [3,] 0.5898411 0.4101589
#>  [4,] 0.5854113 0.4145887
#>  [5,] 0.5891414 0.4108586
#>  [6,] 0.5809058 0.4190942
#>  [7,] 0.5906462 0.4093538
#>  [8,] 0.5864470 0.4135530
#>  [9,] 0.5846761 0.4153239
#> [10,] 0.5915042 0.4084958
#> [11,] 0.5923428 0.4076572
#> [12,] 0.5945316 0.4054684
#> [13,] 0.5891084 0.4108916
#> [14,] 0.5942558 0.4057442
#> [15,] 0.5846633 0.4153367
#> [16,] 0.5857726 0.4142274
#> [17,] 0.5882645 0.4117355
#> [18,] 0.5802858 0.4197142
#> [19,] 0.5863608 0.4136392
#> [20,] 0.5774610 0.4225390
#> [21,] 0.5739734 0.4260266
#> [22,] 0.5832969 0.4167031
#> [23,] 0.5807242 0.4192758
#> [24,] 0.5747185 0.4252815
#> [25,] 0.5889612 0.4110388
#> [26,] 0.5863850 0.4136150
#> [27,] 0.5902736 0.4097264
#> [28,] 0.5759483 0.4240517
#> [29,] 0.5876238 0.4123762
#> [30,] 0.5725844 0.4274156
#> [31,] 0.5844251 0.4155749
#> [32,] 0.5858559 0.4141441
#> [33,] 0.5903794 0.4096206
#> [34,] 0.5859295 0.4140705
#> [35,] 0.5796139 0.4203861
#> [36,] 0.5956911 0.4043089
#> [37,] 0.5852756 0.4147244
#> [38,] 0.5771746 0.4228254
#> [39,] 0.5854195 0.4145805
#> [40,] 0.5772120 0.4227880
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]         [,2]         [,3]      [,4]      [,5]         [,6]
#>  [1,] 2.491009  0.247169385  0.511173791 1.3155655 0.7418848 -0.006552305
#>  [2,] 1.975044 -0.514217590 -0.300443210 0.5063665 0.4102252 -1.083346260
#>  [3,] 2.424248 -0.107494234  0.103149373 1.4707310 0.5093697  0.098046196
#>  [4,] 2.238312  0.481200411  0.986807331 0.9210724 0.9188314  0.473946227
#>  [5,] 2.434734  0.378409968  0.948670021 1.3032773 0.8611855  0.397663831
#>  [6,] 2.186373 -0.290130440 -0.007529001 0.7178539 0.4915491 -0.367488026
#>  [7,] 2.979945  0.733853018  1.258713500 1.2820993 1.1068875  0.204216467
#>  [8,] 2.592587  0.362757489  0.840862903 0.9680130 0.8023254 -0.130856794
#>  [9,] 2.437943  0.331889469  0.884477648 0.9692715 0.8095048 -0.190463069
#> [10,] 2.935666  0.753603755  1.290097971 1.3316883 1.1135178  0.402096843
#> [11,] 2.719047  0.449177608  0.893737842 1.2976639 0.8778022  0.514847647
#> [12,] 2.919036  0.362742924  0.808224698 1.6238285 0.9099949  0.506253684
#> [13,] 2.415832 -0.179090434  0.062036534 1.4370660 0.5012318  0.031388436
#> [14,] 2.690214  0.464353223  0.886343545 1.7376575 0.9157789  0.652658969
#> [15,] 2.371331  0.236669865  0.679885884 0.9246890 0.8584283  0.023910519
#> [16,] 2.145101  0.318712119  0.715956012 1.0530659 0.6913643  0.239717224
#> [17,] 2.578751  0.053202101  0.470431382 0.9777396 0.7164577 -0.018352066
#> [18,] 2.363977 -0.059035094  0.627377638 0.5280773 0.6939998 -0.491307486
#> [19,] 2.632947  0.398870755  0.962937461 0.9803113 0.8935055  0.009679929
#> [20,] 2.241704 -0.292698264  0.344831858 0.4173377 0.6109577 -0.748714523
#> [21,] 1.896538 -0.154392525  0.135172083 0.3771612 0.5771605 -0.534586303
#> [22,] 2.359508 -0.124034783  0.251214934 1.0141903 0.6066965 -0.262285599
#> [23,] 2.184099 -0.405540963  0.110532526 0.5638633 0.4314367 -0.722127534
#> [24,] 2.032833 -0.001173058  0.337467616 0.3871741 0.6058167 -0.816357783
#> [25,] 2.350624  0.126497876  0.372221295 1.4410332 0.6353813  0.060226959
#> [26,] 2.508930 -0.371081950  0.030822501 0.9254331 0.5984215 -0.533221725
#> [27,] 2.683718  0.521155047  1.226464062 1.1446473 1.0270145  0.476419839
#> [28,] 1.887582  0.205822037  0.375736729 0.5730047 0.6939123 -0.259168855
#> [29,] 2.648757  0.523063184  0.950482180 1.0992967 0.9544057  0.049083746
#> [30,] 1.959353 -0.532548772 -0.222689251 0.2920594 0.3902183 -0.960786185
#> [31,] 2.175914  0.168131039  0.548033459 0.9784589 0.7074674  0.032888006
#> [32,] 2.483776 -0.048679303  0.358800267 1.1351046 0.7050256 -0.070066698
#> [33,] 2.726223  0.582757607  1.115421903 1.2381670 1.0566367  0.524151411
#> [34,] 2.093041  0.226247863  0.486758317 1.0336128 0.6468694  0.137917412
#> [35,] 2.185105 -0.455553516 -0.330183470 0.8595332 0.4236683 -0.666479388
#> [36,] 2.960043  0.513156957  1.006290969 1.8984270 1.0016961  0.586653220
#> [37,] 2.752449  0.237354901  0.842109882 0.9976534 0.9112290 -0.376606262
#> [38,] 2.037772  0.370907860  0.530080225 0.6156515 0.7510230 -0.256940713
#> [39,] 2.451735  0.388957762  0.795539709 0.8648540 0.8432660  0.142955151
#> [40,] 1.808207 -0.145026033  0.193339888 0.6653904 0.4708774 -0.272086691
#>            [,7]       [,8]          [,9]    [,10]
#>  [1,] 1.3085878 0.76972937  0.0152448040 1.796713
#>  [2,] 0.4435861 0.24035317 -0.7441077916 1.118452
#>  [3,] 1.1566557 0.28438442 -0.0549455161 1.549212
#>  [4,] 1.3174379 0.99854808  0.2642533878 1.624423
#>  [5,] 1.4350728 0.84459428  0.2819702032 1.773608
#>  [6,] 0.8271732 0.35462629 -0.2546107283 1.376014
#>  [7,] 1.7772358 1.30176150  0.2719506345 2.384598
#>  [8,] 1.2280684 1.07703316 -0.0486985952 2.064843
#>  [9,] 1.3346544 0.97076715 -0.0002565869 1.790498
#> [10,] 1.7931263 1.27222715  0.3733329115 2.334282
#> [11,] 1.6275739 0.94342659  0.1924625418 2.008314
#> [12,] 1.7152697 0.82715525  0.1899832722 2.043510
#> [13,] 1.1602769 0.17755914 -0.1233683792 1.479526
#> [14,] 1.7047649 0.82696794  0.3649557124 1.957232
#> [15,] 1.2813411 0.79087896 -0.0759967380 1.607110
#> [16,] 1.1955939 0.73592655  0.0735212059 1.567625
#> [17,] 1.2886748 0.77174643 -0.2622726755 1.781635
#> [18,] 1.0378864 0.82934309 -0.4344712749 1.584743
#> [19,] 1.5258793 1.02325039  0.1068488699 1.849442
#> [20,] 0.8336078 0.70372533 -0.6481275720 1.410563
#> [21,] 0.6977611 0.57869123 -0.4611535031 1.122847
#> [22,] 0.9413488 0.41777092 -0.1819906347 1.487514
#> [23,] 1.0722289 0.43261634 -0.7387401030 1.307128
#> [24,] 0.7730477 0.67865535 -0.3837421817 1.434908
#> [25,] 1.3369180 0.53192978  0.1413091784 1.626370
#> [26,] 1.2251138 0.42107452 -0.4815136283 1.620129
#> [27,] 1.7148195 1.16732274  0.2594191464 1.971204
#> [28,] 0.7752718 0.70247915 -0.0483106392 1.301638
#> [29,] 1.5084904 1.09061463  0.1952181316 2.050057
#> [30,] 0.4546685 0.36857290 -0.7683752630 1.117917
#> [31,] 1.0352550 0.81052692 -0.0203020313 1.576448
#> [32,] 1.0822581 0.50091494 -0.0580356838 1.597646
#> [33,] 1.6501439 1.13197050  0.3330434326 1.982539
#> [34,] 1.1422753 0.66191373  0.0051083369 1.580072
#> [35,] 0.7148950 0.04621423 -0.5424650215 1.226019
#> [36,] 1.8824908 0.91554038  0.3970221047 2.176201
#> [37,] 1.2720160 0.96774316 -0.1615974398 2.014586
#> [38,] 0.8742510 0.74079057  0.0280209967 1.490034
#> [39,] 1.4065193 0.92247017  0.0460135199 1.696688
#> [40,] 0.7913784 0.35525495 -0.2482685237 1.086979
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1633395 -0.1947751
#>  [2,] 0.1304170 -0.1641617
#>  [3,] 0.1737825 -0.1895260
#>  [4,] 0.1454586 -0.1995690
#>  [5,] 0.1578457 -0.2025713
#>  [6,] 0.1471676 -0.1793252
#>  [7,] 0.1575762 -0.2090613
#>  [8,] 0.1503048 -0.1989917
#>  [9,] 0.1455071 -0.1964922
#> [10,] 0.1596133 -0.2105740
#> [11,] 0.1675025 -0.2061562
#> [12,] 0.1759787 -0.2067523
#> [13,] 0.1723065 -0.1879743
#> [14,] 0.1744028 -0.2071841
#> [15,] 0.1481266 -0.1938201
#> [16,] 0.1513920 -0.1951245
#> [17,] 0.1619201 -0.1948755
#> [18,] 0.1351350 -0.1888116
#> [19,] 0.1488721 -0.2000690
#> [20,] 0.1316340 -0.1807251
#> [21,] 0.1254802 -0.1726011
#> [22,] 0.1508790 -0.1854431
#> [23,] 0.1465825 -0.1791644
#> [24,] 0.1228412 -0.1782878
#> [25,] 0.1670501 -0.1926224
#> [26,] 0.1637738 -0.1852672
#> [27,] 0.1570888 -0.2080079
#> [28,] 0.1266344 -0.1795282
#> [29,] 0.1524235 -0.2017275
#> [30,] 0.1267163 -0.1656872
#> [31,] 0.1499048 -0.1910609
#> [32,] 0.1565223 -0.1903376
#> [33,] 0.1586653 -0.2068688
#> [34,] 0.1551719 -0.1919913
#> [35,] 0.1500127 -0.1711759
#> [36,] 0.1770158 -0.2105271
#> [37,] 0.1470949 -0.1973740
#> [38,] 0.1267920 -0.1843937
#> [39,] 0.1481090 -0.1969524
#> [40,] 0.1349431 -0.1763957
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4114160 -0.4114160
#>  [2,] -0.5731167  0.5731167
#>  [3,]  0.4101589 -0.4101589
#>  [4,] -0.5854113  0.5854113
#>  [5,] -0.5891414  0.5891414
#>  [6,]  0.4190942 -0.4190942
#>  [7,] -0.5906462  0.5906462
#>  [8,]  0.4135530 -0.4135530
#>  [9,] -0.5846761  0.5846761
#> [10,] -0.5915042  0.5915042
#> [11,]  0.4076572 -0.4076572
#> [12,]  0.4054684 -0.4054684
#> [13,]  0.4108916 -0.4108916
#> [14,] -0.5942558  0.5942558
#> [15,]  0.4153367 -0.4153367
#> [16,] -0.5857726  0.5857726
#> [17,]  0.4117355 -0.4117355
#> [18,]  0.4197142 -0.4197142
#> [19,] -0.5863608  0.5863608
#> [20,]  0.4225390 -0.4225390
#> [21,]  0.4260266 -0.4260266
#> [22,] -0.5832969  0.5832969
#> [23,]  0.4192758 -0.4192758
#> [24,] -0.5747185  0.5747185
#> [25,]  0.4110388 -0.4110388
#> [26,]  0.4136150 -0.4136150
#> [27,]  0.4097264 -0.4097264
#> [28,] -0.5759483  0.5759483
#> [29,] -0.5876238  0.5876238
#> [30,] -0.5725844  0.5725844
#> [31,] -0.5844251  0.5844251
#> [32,] -0.5858559  0.5858559
#> [33,]  0.4096206 -0.4096206
#> [34,]  0.4140705 -0.4140705
#> [35,] -0.5796139  0.5796139
#> [36,] -0.5956911  0.5956911
#> [37,] -0.5852756  0.5852756
#> [38,] -0.5771746  0.5771746
#> [39,] -0.5854195  0.5854195
#> [40,]  0.4227880 -0.4227880
#> 
#> $L
#> [1] 0.7008504 0.7261480 0.7384392 0.7437031 0.7413689 0.7150031
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5942029