Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12           V13
#>  [1,] -0.2446331 -0.03480688 -0.063963873 -0.105451977 -4.586946e-02
#>  [2,] -0.3102883 -0.08642026 -0.142116364  0.034412758 -7.316875e-02
#>  [3,] -0.2398554 -0.09849548 -0.195650763 -0.188551000 -3.960538e-02
#>  [4,] -0.2449200 -0.06801183  0.035095416 -0.052148818 -1.453316e-05
#>  [5,] -0.3416733 -0.18736360 -0.071964871 -0.068748612 -8.881706e-02
#>  [6,] -0.2484566 -0.07713736  0.006158499  0.054893352 -6.878559e-02
#>  [7,] -0.2863221 -0.09740291 -0.032855396 -0.140612688 -4.821797e-02
#>  [8,] -0.3196545 -0.02083171 -0.066028184 -0.000326355 -7.002040e-03
#>  [9,] -0.3920322 -0.18202291 -0.167073957 -0.105358511 -8.114892e-02
#> [10,] -0.2035015 -0.12514179 -0.030277105 -0.081184447 -8.042997e-02
#>                V14          V15          V16          V17         V18
#>  [1,] -0.029598415  0.011993914  0.107367280  0.063588526  0.13606986
#>  [2,]  0.027736798 -0.044935827  0.113350176  0.005387170  0.11072790
#>  [3,] -0.044047887 -0.163001549 -0.069776264 -0.131226249 -0.14208157
#>  [4,] -0.003870059 -0.085613593  0.039545584  0.035098307  0.09902062
#>  [5,] -0.159019988 -0.154832885 -0.006168604  0.020698477 -0.05064637
#>  [6,]  0.013839330  0.019822528  0.036839332  0.117479249  0.10280545
#>  [7,] -0.161755077  0.046474295  0.008520426  0.005899479  0.11245081
#>  [8,] -0.044454624  0.071825545  0.091577927  0.156525797  0.13783805
#>  [9,]  0.012393421 -0.005299494 -0.127771494 -0.108242347 -0.11493970
#> [10,]  0.018949142 -0.018208906 -0.084667826 -0.091617830 -0.05549553
#>                V19         V2         V20         V21         V22         V23
#>  [1,]  0.108561685 -0.1857713  0.12930951  0.17763714  0.10351180  0.04316453
#>  [2,]  0.010461054 -0.2896782  0.09641223  0.08163107  0.15053996  0.26886328
#>  [3,] -0.060492675 -0.2937529 -0.10565535  0.09292542 -0.03285172  0.10173260
#>  [4,]  0.191201615 -0.1386692  0.10457030  0.14079699  0.19151223  0.23791998
#>  [5,]  0.010342972 -0.3537332  0.09946523  0.09648235  0.17736879  0.16347623
#>  [6,]  0.070222848 -0.3051489  0.22888040  0.22871274  0.29727140  0.17925118
#>  [7,] -0.014955039 -0.2485370  0.07370019  0.04638635  0.07986640  0.20975158
#>  [8,]  0.072123933 -0.2939048  0.05495784  0.09791185  0.18885656  0.15419667
#>  [9,]  0.079010681 -0.2135546  0.05445312 -0.01461183 -0.11451208 -0.05704017
#> [10,]  0.002870191 -0.2138709  0.02625650  0.09164493  0.12374562  0.06014266
#>              V24        V25        V26          V27           V28          V29
#>  [1,] 0.08659284 0.07910204 0.03740442  0.069575499 -0.0017372411 -0.015474795
#>  [2,] 0.13919902 0.29654459 0.25203587  0.241365271  0.3176409905  0.104411672
#>  [3,] 0.08036674 0.08064166 0.20188842  0.110080472  0.0903353642  0.013373007
#>  [4,] 0.11638994 0.08537073 0.11630075  0.126403405  0.2267816709  0.222530958
#>  [5,] 0.21706882 0.17663860 0.20146131  0.322627785  0.2197402343  0.225170004
#>  [6,] 0.24017331 0.25829392 0.25052641  0.386155645  0.2611008177  0.127061783
#>  [7,] 0.31989299 0.30335216 0.20116478  0.277731110  0.1583224777  0.149834197
#>  [8,] 0.09261963 0.05602859 0.13480397 -0.006016157  0.0841664150 -0.002306045
#>  [9,] 0.02104620 0.12043309 0.07479671  0.070001909 -0.0008305495  0.091398814
#> [10,] 0.20372515 0.16596029 0.31957800  0.183103943  0.1266798466  0.196241174
#>               V3          V30         V31          V32          V33
#>  [1,] -0.3223412  0.002331725  0.07434954 -0.013741586 -0.100780362
#>  [2,] -0.2981544  0.145396497  0.00687569 -0.020932669  0.031261462
#>  [3,] -0.1534070 -0.034637529  0.02575233 -0.070281950  0.009883466
#>  [4,] -0.2816162  0.086754876 -0.01820723 -0.087381839 -0.084046954
#>  [5,] -0.3536063  0.158242478  0.11214784  0.084703912  0.010467373
#>  [6,] -0.3752180  0.174527278  0.16356852  0.051342816 -0.003148371
#>  [7,] -0.3041879  0.166081261  0.12184540 -0.087499040 -0.084047552
#>  [8,] -0.2349167  0.053139618  0.05354052  0.007756034  0.055441771
#>  [9,] -0.2744277  0.101334176 -0.06792735  0.037567145 -0.049153826
#> [10,] -0.1832731  0.020459314  0.06037760  0.145006158  0.065698173
#>                V34          V35          V36         V37          V38
#>  [1,] -0.019407457 -0.054030935 -0.018196716 -0.04434666 -0.064217282
#>  [2,] -0.080482905  0.002103917 -0.010509079  0.07831445  0.028562970
#>  [3,] -0.117908596 -0.082579916  0.019750938  0.09419911  0.001385183
#>  [4,]  0.025376298 -0.091118347 -0.106585353 -0.11416560 -0.077512265
#>  [5,] -0.113165484 -0.013766907 -0.073766364 -0.10287018 -0.136803827
#>  [6,] -0.004572410  0.042642675  0.005618454 -0.02198151 -0.024157295
#>  [7,] -0.048136139 -0.049413822 -0.130242229 -0.04964462 -0.124927657
#>  [8,]  0.003344398  0.019068662 -0.073392336 -0.02250658 -0.141202136
#>  [9,]  0.007407335 -0.132709618  0.045800536 -0.11331490 -0.097785470
#> [10,] -0.030367135 -0.086589350  0.003865038 -0.03621057 -0.018422960
#>                V39         V4         V40         V41         V42           V43
#>  [1,] -0.067737495 -0.2097849 -0.01238792 -0.15057948 -0.09930766 -0.2272510754
#>  [2,] -0.009633849 -0.2935812 -0.06518471  0.04606689 -0.10321551 -0.0005391217
#>  [3,] -0.013907075 -0.1434091 -0.03992619 -0.17547980 -0.06075702 -0.1515728883
#>  [4,] -0.074474311 -0.1637516 -0.11907943 -0.12972058 -0.09442561 -0.1043446733
#>  [5,] -0.075310675 -0.4116245 -0.12465668 -0.16288886 -0.15677986 -0.0605341609
#>  [6,] -0.078455583 -0.2549759 -0.04376984  0.05085123 -0.11765434 -0.0246089414
#>  [7,]  0.060446335 -0.3248511 -0.03711294 -0.06858408 -0.06225470 -0.0638550370
#>  [8,] -0.122040650 -0.2177282 -0.12430368 -0.05625403 -0.05454134 -0.0971004430
#>  [9,] -0.130191893 -0.2699817 -0.10779894 -0.06580463 -0.13534106 -0.1615660888
#> [10,] -0.081674754 -0.2501961 -0.06343087 -0.14169490  0.02858171 -0.1170766705
#>               V44         V45         V46         V47        V48        V49
#>  [1,] -0.17477656 -0.20612347 -0.21385263 -0.15370835 -0.1878164 -0.3185310
#>  [2,] -0.03101206 -0.12008670 -0.07557084 -0.26000247 -0.1849913 -0.2832553
#>  [3,] -0.06840155 -0.18005008 -0.05689563 -0.19806377 -0.1768752 -0.2489152
#>  [4,] -0.09900691 -0.10186350 -0.10484599 -0.22060645 -0.2278967 -0.1391322
#>  [5,] -0.07165405 -0.21480147 -0.22403393 -0.15707913 -0.2227556 -0.3223681
#>  [6,] -0.12113912 -0.10725477 -0.13159007 -0.10667745 -0.2808609 -0.1874647
#>  [7,] -0.09936856 -0.09222052 -0.24866010 -0.20900158 -0.1336287 -0.2565810
#>  [8,] -0.09978109 -0.19706251 -0.25593831 -0.16387609 -0.2344916 -0.3474943
#>  [9,] -0.22036657 -0.13143160 -0.16410743 -0.22846041 -0.2808968 -0.1739925
#> [10,] -0.03367759 -0.10390643 -0.15435970 -0.06995964 -0.1106511 -0.3129644
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.2771946 -0.2255395 -0.3002099 -0.3013113 -0.2783748 -0.3419760
#>  [2,] -0.2379538 -0.3199824 -0.3564297 -0.4529054 -0.4872624 -0.3244121
#>  [3,] -0.1818436 -0.2595064 -0.2946604 -0.3363685 -0.2682013 -0.1849116
#>  [4,] -0.1487511 -0.1578613 -0.3012704 -0.3456590 -0.2736047 -0.2010111
#>  [5,] -0.3207378 -0.4515967 -0.4348277 -0.3709838 -0.3917399 -0.4759159
#>  [6,] -0.3276308 -0.3044366 -0.3217275 -0.4093902 -0.4413284 -0.4054367
#>  [7,] -0.1782123 -0.2268256 -0.1786072 -0.1904112 -0.2135090 -0.3380744
#>  [8,] -0.1727277 -0.4097437 -0.3892039 -0.3204907 -0.3442210 -0.2391900
#>  [9,] -0.1527026 -0.3133758 -0.3629037 -0.2776249 -0.2374439 -0.2530362
#> [10,] -0.2960496 -0.3602350 -0.1806829 -0.2207723 -0.2647237 -0.3631111
#>              V55        V56        V57        V58        V59         V6
#>  [1,] -0.2637493 -0.3861710 -0.3358707 -0.2434106 -0.2602032 -0.1815827
#>  [2,] -0.3731349 -0.3482013 -0.3797689 -0.4455078 -0.4429110 -0.1506193
#>  [3,] -0.3060376 -0.1833828 -0.2800467 -0.2534712 -0.2143630 -0.1938065
#>  [4,] -0.3213202 -0.3543521 -0.3377995 -0.3098958 -0.3474644 -0.1045125
#>  [5,] -0.4787776 -0.4063236 -0.4346838 -0.3786387 -0.3560625 -0.3111224
#>  [6,] -0.3348750 -0.4200046 -0.3986767 -0.4122474 -0.2965034 -0.2114160
#>  [7,] -0.3443941 -0.1918728 -0.3400640 -0.1942374 -0.3283821 -0.2671181
#>  [8,] -0.3061264 -0.2321940 -0.3226267 -0.4065545 -0.3032925 -0.1174679
#>  [9,] -0.3134269 -0.3064081 -0.3051719 -0.2456343 -0.3967414 -0.2421727
#> [10,] -0.3286919 -0.3118407 -0.2037596 -0.2878539 -0.4077957 -0.2065115
#>              V60          V7          V8          V9
#>  [1,] -0.3033975 -0.20845677 -0.16667839 -0.17694442
#>  [2,] -0.4251682 -0.15130696 -0.19725398 -0.10962783
#>  [3,] -0.2681576 -0.08382325 -0.23782254 -0.07282105
#>  [4,] -0.3029769 -0.11245764 -0.12518138 -0.08779339
#>  [5,] -0.3572756 -0.15193228 -0.23767498 -0.08965569
#>  [6,] -0.3168990 -0.08475430 -0.21328970 -0.11574727
#>  [7,] -0.2592309 -0.08273253 -0.06632512 -0.15585200
#>  [8,] -0.3923878 -0.16251125 -0.08650900 -0.09686632
#>  [9,] -0.3541479 -0.24105371 -0.09976488 -0.08308487
#> [10,] -0.2893366 -0.26100062 -0.14029582 -0.20247562
#> 
#> $W[[2]]
#>            [,1]        [,2]       [,3]        [,4]       [,5]         [,6]
#> [1,] -0.1192532 0.004182458 -0.1561204  0.08584918 0.02695443  0.013436629
#> [2,]  0.1195910 0.122026281  0.1620052 -0.05623043 0.02796964 -0.008602059
#>            [,7]         [,8]        [,9]       [,10]
#> [1,] 0.07251888 -0.070561939 -0.04471163 -0.07740027
#> [2,] 0.09565915  0.005621578  0.11894580  0.06567304
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  3.344828e-04  0.0020189971  0.0022602903  0.0017107448  1.455239e-03
#>  [2,]  1.062806e-04  0.0005033012  0.0005081470  0.0003361322  3.091997e-04
#>  [3,]  3.780653e-04  0.0023461377  0.0026297915  0.0020220372  1.675051e-03
#>  [4,] -2.118870e-04 -0.0013129931 -0.0014998069 -0.0010620290 -8.058228e-04
#>  [5,] -5.670670e-05 -0.0004374497 -0.0005585477 -0.0004288367 -3.316975e-04
#>  [6,] -3.344779e-05 -0.0002202959 -0.0002776706 -0.0001786879 -1.235567e-04
#>  [7,] -6.411450e-06 -0.0001359398 -0.0002084174 -0.0001633309 -9.879518e-05
#>  [8,]  8.496973e-05  0.0004223662  0.0004266334  0.0003086547  2.938633e-04
#>  [9,]  1.566082e-04  0.0008740122  0.0009483848  0.0007204695  6.404122e-04
#> [10,]  1.663835e-04  0.0008504935  0.0009003136  0.0006375314  5.904840e-04
#>                 V14           V15           V16           V17          V18
#>  [1,]  1.589174e-03  1.539217e-03  1.345806e-03  7.212666e-04 1.912284e-04
#>  [2,]  4.675160e-04  5.888872e-04  6.037947e-04  4.368754e-04 2.353659e-04
#>  [3,]  1.804008e-03  1.696614e-03  1.451652e-03  7.320146e-04 9.336923e-05
#>  [4,] -8.123476e-04 -7.009364e-04 -4.844272e-04 -2.627866e-05 2.688823e-04
#>  [5,] -2.531659e-04 -1.302315e-04 -2.340188e-05  8.722115e-05 7.311943e-05
#>  [6,] -1.204450e-04 -9.291172e-05 -4.652162e-05  2.067747e-05 4.715078e-05
#>  [7,] -3.135328e-06  1.051684e-04  1.859990e-04  2.196202e-04 1.615438e-04
#>  [8,]  4.182552e-04  4.994003e-04  5.150573e-04  3.731985e-04 1.796026e-04
#>  [9,]  7.922421e-04  8.415462e-04  8.035937e-04  5.133584e-04 1.613687e-04
#> [10,]  8.083531e-04  9.210205e-04  9.143331e-04  6.334345e-04 3.250253e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  1.710804e-03  3.619159e-04  2.550802e-03  2.395113e-03  9.312879e-04
#>  [2,]  5.801714e-04  1.204058e-04  9.047498e-04  8.664644e-04  4.449611e-04
#>  [3,]  1.866440e-03  4.077538e-04  2.758912e-03  2.522031e-03  8.377026e-04
#>  [4,] -7.452844e-04 -2.089451e-04 -1.315570e-03 -1.180403e-03 -2.383049e-04
#>  [5,] -2.424899e-04 -4.466133e-05 -3.872091e-04 -3.688469e-04 -1.695109e-04
#>  [6,] -9.854993e-05 -2.603134e-05 -2.039586e-04 -1.885381e-04 -5.532421e-05
#>  [7,]  6.012073e-05  6.066902e-06  5.644962e-05  5.507102e-05  6.883114e-05
#>  [8,]  4.894674e-04  1.014512e-04  7.189844e-04  6.741668e-04  3.048807e-04
#>  [9,]  8.378190e-04  1.777059e-04  1.199696e-03  1.069602e-03  3.108420e-04
#> [10,]  9.571358e-04  1.872461e-04  1.437251e-03  1.356030e-03  6.511629e-04
#>                 V23           V24           V25           V26           V27
#>  [1,]  1.151805e-04 -6.761868e-05 -0.0005741792 -0.0002550966 -1.955679e-05
#>  [2,]  2.371585e-04  2.516883e-04  0.0001973490  0.0003219072  3.203625e-04
#>  [3,] -8.080949e-05 -3.294444e-04 -0.0009570582 -0.0006196249 -2.438424e-04
#>  [4,]  3.624936e-04  5.351678e-04  0.0008785103  0.0007534539  6.716691e-04
#>  [5,]  1.991887e-05  1.303912e-04  0.0002926334  0.0003217268  3.232927e-04
#>  [6,]  8.157293e-05  1.305175e-04  0.0001868121  0.0001999265  2.194800e-04
#>  [7,]  1.388295e-04  2.105637e-04  0.0002978620  0.0003606327  3.637466e-04
#>  [8,]  1.486527e-04  1.641964e-04  0.0001150140  0.0002542912  3.093866e-04
#>  [9,] -5.160525e-05 -1.158360e-04 -0.0003404576 -0.0001587735 -1.074888e-06
#> [10,]  3.043849e-04  2.765270e-04  0.0001104018  0.0003104424  3.418885e-04
#>                V28          V29            V3          V30           V31
#>  [1,] 0.0002804165 2.069268e-04  1.589980e-04 0.0003219328 -4.877720e-04
#>  [2,] 0.0004107121 4.440799e-04  5.451228e-05 0.0005795756  3.550194e-04
#>  [3,] 0.0001212811 5.353643e-05  1.685911e-04 0.0001851813 -7.079023e-04
#>  [4,] 0.0005063686 5.695201e-04 -9.692155e-05 0.0004410987  9.443486e-04
#>  [5,] 0.0003401916 4.213691e-04 -3.897277e-05 0.0004625059  6.070515e-04
#>  [6,] 0.0002570609 3.034697e-04 -2.037824e-05 0.0002822599  3.619921e-04
#>  [7,] 0.0003980991 4.631863e-04 -8.015795e-06 0.0005142557  5.340027e-04
#>  [8,] 0.0004045692 4.318452e-04  3.874197e-05 0.0005154517  3.138314e-04
#>  [9,] 0.0001667477 2.146293e-04  6.216596e-05 0.0003429960 -1.591745e-05
#> [10,] 0.0004633708 5.158015e-04  7.707437e-05 0.0006724664  2.949432e-04
#>                 V32           V33           V34           V35           V36
#>  [1,] -4.050053e-04 -1.378177e-03 -0.0019140598 -2.385475e-03 -3.276906e-03
#>  [2,]  3.110536e-04  1.510988e-04  0.0001706410  2.127793e-04  2.434915e-05
#>  [3,] -5.934817e-04 -1.887794e-03 -0.0026332733 -3.204002e-03 -4.207320e-03
#>  [4,]  7.534257e-04  1.359513e-03  0.0017114184  2.052561e-03  2.782165e-03
#>  [5,]  4.644811e-04  6.109871e-04  0.0007629990  9.431289e-04  1.238115e-03
#>  [6,]  2.960345e-04  3.769588e-04  0.0004496680  5.154880e-04  6.655933e-04
#>  [7,]  4.217037e-04  4.680165e-04  0.0005731109  6.936234e-04  8.251724e-04
#>  [8,]  2.428329e-04  2.950821e-05 -0.0000139340 -2.378606e-05 -1.707759e-04
#>  [9,] -3.622644e-05 -5.640734e-04 -0.0008118674 -9.616345e-04 -1.302302e-03
#> [10,]  2.753662e-04 -7.604271e-05 -0.0001473930 -1.940739e-04 -5.474047e-04
#>                 V37           V38           V39            V4           V40
#>  [1,] -2.489201e-03 -7.336467e-04 -4.359194e-05  2.762087e-04 -8.632922e-04
#>  [2,]  6.981194e-05  3.294246e-04  3.141468e-04  7.810466e-05  1.424760e-05
#>  [3,] -3.131874e-03 -9.930441e-04 -1.875854e-04  3.048018e-04 -1.137517e-03
#>  [4,]  2.317987e-03  1.084859e-03  5.687326e-04 -1.687436e-04  1.052536e-03
#>  [5,]  1.102403e-03  6.617071e-04  4.077716e-04 -6.035087e-05  4.705693e-04
#>  [6,]  5.955629e-04  3.872276e-04  2.914324e-04 -3.132245e-05  3.366536e-04
#>  [7,]  7.459499e-04  5.400592e-04  3.792890e-04 -1.298772e-05  3.271733e-04
#>  [8,] -5.102862e-05  2.301083e-04  2.654716e-04  6.537840e-05  7.109773e-06
#>  [9,] -8.760111e-04 -8.391621e-05  1.233870e-04  1.179524e-04 -3.358818e-04
#> [10,] -3.576704e-04  2.405068e-04  3.364350e-04  1.271402e-04 -1.412637e-04
#>                 V41          V42           V43           V44           V45
#>  [1,] -6.240135e-04 8.564555e-05  8.529976e-04  1.378565e-03  1.768249e-03
#>  [2,]  5.212361e-05 1.817750e-04  3.137214e-04  4.358651e-04  5.492650e-04
#>  [3,] -8.207072e-04 2.896897e-05  9.578484e-04  1.613531e-03  2.016664e-03
#>  [4,]  8.226494e-04 2.803408e-04 -3.503613e-04 -7.619634e-04 -1.099575e-03
#>  [5,]  3.935691e-04 1.976632e-04 -4.840511e-05 -1.521344e-04 -2.472365e-04
#>  [6,]  2.872494e-04 1.813111e-04  3.751104e-05 -4.522233e-05 -1.157815e-04
#>  [7,]  2.829142e-04 1.906702e-04  7.622106e-05  3.990446e-05  1.250406e-06
#>  [8,]  4.942009e-05 1.691650e-04  2.922770e-04  4.022057e-04  4.800864e-04
#>  [9,] -1.934018e-04 1.184899e-04  4.335415e-04  6.860585e-04  8.438806e-04
#> [10,] -5.105804e-05 2.070035e-04  4.951514e-04  7.164302e-04  8.761492e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.331788e-03  8.844620e-04  7.118081e-04  4.223677e-04  2.765506e-04
#>  [2,]  4.378731e-04  2.577492e-04  1.841533e-04  1.011835e-04  7.946337e-05
#>  [3,]  1.472212e-03  1.008035e-03  8.203271e-04  4.837840e-04  2.972605e-04
#>  [4,] -8.341331e-04 -5.121485e-04 -4.144651e-04 -2.428408e-04 -1.508074e-04
#>  [5,] -1.779137e-04 -1.052023e-04 -9.822091e-05 -6.111738e-05 -5.220807e-05
#>  [6,] -8.451306e-05 -3.877010e-05 -3.818577e-05 -2.037324e-05 -1.999534e-05
#>  [7,]  8.394045e-06  7.292640e-06 -8.784489e-06 -8.915664e-06 -5.899321e-06
#>  [8,]  3.704582e-04  2.396149e-04  1.787966e-04  1.017991e-04  6.830524e-05
#>  [9,]  6.189686e-04  4.229993e-04  3.360509e-04  1.964578e-04  1.107995e-04
#> [10,]  6.734627e-04  4.135854e-04  3.107620e-04  1.784950e-04  1.305057e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  4.435973e-05  1.361771e-04  1.225562e-04  3.811199e-05  5.992945e-05
#>  [2,]  2.093478e-05  4.161244e-05  3.840078e-05  1.642732e-05  2.226404e-05
#>  [3,]  3.500988e-05  1.511778e-04  1.385015e-04  4.029716e-05  6.445024e-05
#>  [4,] -1.714830e-05 -7.963973e-05 -7.942186e-05 -1.541803e-05 -3.017848e-05
#>  [5,] -2.211506e-06 -1.947272e-05 -2.193210e-05 -2.727874e-06 -6.019771e-06
#>  [6,]  2.349658e-06 -8.489739e-06 -1.200980e-05  3.885581e-07 -2.057636e-06
#>  [7,]  5.799498e-06 -2.881961e-07 -3.028035e-06  4.467731e-06  3.458033e-06
#>  [8,]  1.724315e-05  3.628164e-05  3.159226e-05  1.394511e-05  1.856961e-05
#>  [9,]  1.664251e-05  6.438098e-05  5.776346e-05  1.979927e-05  2.975569e-05
#> [10,]  2.853460e-05  6.705196e-05  6.027936e-05  2.560809e-05  3.463624e-05
#>                 V55           V56          V57           V58           V59
#>  [1,]  5.290619e-05  3.780095e-05 8.328099e-06  5.756741e-05  4.446307e-05
#>  [2,]  2.222364e-05  1.262988e-05 1.028184e-05  1.839174e-05  1.507530e-05
#>  [3,]  5.547728e-05  4.236669e-05 7.814082e-06  6.397518e-05  5.010995e-05
#>  [4,] -2.432688e-05 -1.821248e-05 1.014451e-06 -3.254256e-05 -2.058123e-05
#>  [5,] -2.616765e-06 -4.628537e-06 3.415332e-06 -9.627686e-06 -4.542180e-06
#>  [6,] -9.594793e-07 -1.311951e-06 7.840642e-07 -3.770503e-06 -7.502774e-07
#>  [7,]  5.602254e-06  1.808179e-06 5.560003e-06 -5.039819e-08  2.421592e-06
#>  [8,]  1.853606e-05  1.168557e-05 7.349261e-06  1.535161e-05  1.332750e-05
#>  [9,]  2.949991e-05  1.927092e-05 6.620318e-06  2.736505e-05  2.301638e-05
#> [10,]  3.436283e-05  2.147230e-05 1.178154e-05  3.030405e-05  2.505426e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.189123e-04  1.948842e-05  6.183389e-04  1.071433e-03  1.893204e-03
#>  [2,]  1.068193e-04  8.770189e-06  2.027028e-04  3.188836e-04  5.134515e-04
#>  [3,]  3.385589e-04  1.823189e-05  6.846656e-04  1.228572e-03  2.214477e-03
#>  [4,] -1.603352e-04 -7.920312e-06 -3.232583e-04 -6.458118e-04 -1.243328e-03
#>  [5,] -5.636716e-05 -1.616105e-06 -8.382955e-05 -1.830716e-04 -3.768533e-04
#>  [6,] -2.140204e-05  6.351188e-10 -3.610184e-05 -9.184558e-05 -2.079028e-04
#>  [7,] -1.565867e-07  1.942893e-06  1.158818e-05 -2.201985e-05 -9.472749e-05
#>  [8,]  8.568955e-05  6.726967e-06  1.723622e-04  2.676411e-04  4.322783e-04
#>  [9,]  1.236073e-04  8.246805e-06  2.887105e-04  4.958663e-04  8.511071e-04
#> [10,]  1.592918e-04  1.251362e-05  3.238674e-04  5.174992e-04  8.447746e-04
#> 
#> $vW[[2]]
#>              [,1]        [,2]        [,3]         [,4]        [,5]         [,6]
#> [1,]  0.009783756  0.01177565  0.01984773  0.003974072  0.01532555  0.007695293
#> [2,] -0.009783756 -0.01177565 -0.01984773 -0.003974072 -0.01532555 -0.007695293
#>             [,7]         [,8]        [,9]       [,10]
#> [1,]  0.01245875  0.009961578  0.01744423  0.01454791
#> [2,] -0.01245875 -0.009961578 -0.01744423 -0.01454791
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3247516 0.7894378 0.2278392 0.3933058 0.6231979 0.8797865 0.2282273
#>  [8] 0.2709791 0.4294940 0.4592214
#> 
#> $B[[2]]
#> [1] -0.01512384 -0.02408250
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0001733735  0.0007895773 -0.0002451571  0.0009537114  0.0006135478
#>  [6]  0.0004207956  0.0007087051  0.0006239886  0.0002774343  0.0009083948
#> 
#> $vB[[2]]
#> [1]  0.007036414 -0.007036414
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#>  [2,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#>  [3,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [4,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#>  [5,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#>  [6,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#>  [7,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#>  [8,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#>  [9,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [10,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [11,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [12,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [13,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [14,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [15,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [16,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [17,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [18,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [19,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [20,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [21,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [22,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [23,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [24,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [25,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [26,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [27,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [28,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [29,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [30,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [31,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [32,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [33,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [34,] 0.0209 0.1853 0.1748 0.1556 0.1476 0.1378 0.2584 0.3827 0.4784 0.5360
#> [35,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [36,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [37,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [38,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [39,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [40,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#>  [2,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#>  [3,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [4,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#>  [5,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#>  [6,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#>  [7,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#>  [8,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#>  [9,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [10,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [11,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [12,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [13,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [14,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [15,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [16,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [17,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [18,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [19,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [20,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [21,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [22,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [23,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [24,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [25,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [26,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [27,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [28,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [29,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [30,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [31,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [32,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [33,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [34,] 0.6192 0.0278 0.7912 0.9264 1.0000 0.9080 0.7435 0.5557 0.3172 0.1295
#> [35,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [36,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [37,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [38,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [39,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [40,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#>  [2,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#>  [3,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [4,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#>  [5,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#>  [6,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#>  [7,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#>  [8,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#>  [9,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [10,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [11,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [12,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [13,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [14,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [15,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [16,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [17,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [18,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [19,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [20,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [21,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [22,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [23,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [24,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [25,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [26,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [27,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [28,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [29,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [30,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [31,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [32,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [33,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [34,] 0.0598 0.2722 0.0115 0.3616 0.3293 0.4855 0.3936 0.1845 0.0342 0.2489
#> [35,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [36,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [37,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [38,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [39,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [40,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#>  [2,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#>  [3,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [4,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#>  [5,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#>  [6,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#>  [7,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#>  [8,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#>  [9,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [10,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [11,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [12,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [13,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [14,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [15,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [16,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [17,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [18,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [19,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [20,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [21,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [22,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [23,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [24,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [25,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [26,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [27,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [28,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [29,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [30,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [31,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [32,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [33,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [34,] 0.3837 0.3514 0.2654 0.0445 0.1760 0.1599 0.0866 0.0590 0.0813 0.0492
#> [35,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [36,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [37,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [38,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [39,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [40,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#>  [2,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#>  [3,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [4,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#>  [5,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#>  [6,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#>  [7,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#>  [8,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#>  [9,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [10,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [11,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [12,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [13,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [14,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [15,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [16,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [17,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [18,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [19,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [20,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [21,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [22,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [23,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [24,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [25,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [26,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [27,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [28,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [29,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [30,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [31,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [32,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [33,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [34,] 0.0417 0.0495 0.0367 0.0115 0.0427 0.0118 0.0133 0.0096 0.0014 0.0049
#> [35,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [36,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [37,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [38,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [39,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [40,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#>  [2,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#>  [3,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [4,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#>  [5,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#>  [6,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#>  [7,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#>  [8,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#>  [9,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [10,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [11,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [12,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [13,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [14,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [15,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [16,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [17,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [18,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [19,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [20,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [21,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [22,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [23,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [24,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [25,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [26,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [27,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [28,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [29,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [30,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [31,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [32,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [33,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [34,] 0.0039 0.0029 0.0078 0.0047 0.0021 0.0766 0.0011 0.1458 0.1430 0.1894
#> [35,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [36,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [37,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [38,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [39,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [40,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6227523 0.8627758 0.4692898 0.7162916 0.7289678 0.9171889 0.7211047
#>  [2,] 0.5369402 0.8239813 0.2946135 0.6805291 0.6192769 0.9067608 0.6241772
#>  [3,] 0.4442599 0.7728506 0.3297640 0.5309204 0.5485240 0.8397329 0.5226730
#>  [4,] 0.5759669 0.8780368 0.4334496 0.7054025 0.7364952 0.9338393 0.7005472
#>  [5,] 0.5783374 0.9069428 0.4653666 0.7233753 0.7838507 0.9466611 0.7540871
#>  [6,] 0.4182739 0.8082881 0.3489037 0.6314991 0.6629634 0.8840337 0.6316647
#>  [7,] 0.6322908 0.8727099 0.3860181 0.7768193 0.7118814 0.9333833 0.7476258
#>  [8,] 0.3076454 0.6853692 0.3167367 0.4114107 0.4097095 0.7256885 0.4277482
#>  [9,] 0.5568628 0.8726962 0.4843548 0.7235887 0.7900454 0.9236093 0.7481713
#> [10,] 0.6703225 0.9246022 0.5036116 0.8307612 0.8507325 0.9611020 0.8334686
#> [11,] 0.5722857 0.8830034 0.5305952 0.7539968 0.8199747 0.9271852 0.7796234
#> [12,] 0.6624310 0.8075501 0.3507039 0.7132757 0.6407478 0.8901800 0.6251339
#> [13,] 0.5822478 0.8797870 0.4452469 0.7098259 0.7260147 0.9252391 0.7402663
#> [14,] 0.3999522 0.8399632 0.4169872 0.5931971 0.6973030 0.8895049 0.6467646
#> [15,] 0.5610850 0.8894048 0.4929450 0.7270672 0.7891602 0.9366183 0.7506473
#> [16,] 0.4210099 0.7297883 0.3177493 0.5272998 0.4834216 0.8178017 0.4631485
#> [17,] 0.6192020 0.7912754 0.3235077 0.6736861 0.5798250 0.8779526 0.5816921
#> [18,] 0.6087153 0.8616244 0.4372669 0.7432396 0.7663604 0.9249719 0.7178125
#> [19,] 0.6443824 0.9191930 0.6038235 0.7943101 0.8712945 0.9538933 0.8256366
#> [20,] 0.6138172 0.8631835 0.4286630 0.7264264 0.7245391 0.9184977 0.7051337
#> [21,] 0.5947252 0.8855015 0.5687466 0.7489454 0.8114642 0.9193210 0.7828368
#> [22,] 0.5621195 0.8366218 0.4275284 0.6441900 0.6181040 0.8823548 0.6233033
#> [23,] 0.3368847 0.7217186 0.2949419 0.5241041 0.5400106 0.8112932 0.5049336
#> [24,] 0.6249962 0.8934826 0.4896094 0.7729526 0.8167021 0.9364026 0.7878611
#> [25,] 0.3686264 0.8281320 0.4158055 0.5577638 0.6702215 0.8775983 0.6301158
#> [26,] 0.5747694 0.8836336 0.4790946 0.7562030 0.8128507 0.9374751 0.7729301
#> [27,] 0.6215873 0.8667461 0.4007326 0.7366668 0.7002892 0.9253797 0.7187850
#> [28,] 0.4370305 0.6989594 0.3309324 0.5106498 0.4709979 0.7986046 0.4519760
#> [29,] 0.5899710 0.9013440 0.5649093 0.7449554 0.8329280 0.9404825 0.7813953
#> [30,] 0.6402027 0.9234064 0.5962883 0.7991142 0.8782001 0.9579893 0.8305236
#> [31,] 0.4898819 0.8567024 0.4518094 0.6423461 0.7164563 0.9062222 0.6904524
#> [32,] 0.6563097 0.8760139 0.4505396 0.7892335 0.7657904 0.9262347 0.7627439
#> [33,] 0.4537965 0.7505667 0.4362112 0.5055678 0.5817776 0.8054231 0.5165546
#> [34,] 0.6664497 0.8466365 0.4800225 0.7466470 0.7534713 0.9093149 0.7100341
#> [35,] 0.3840292 0.7439711 0.3192067 0.5171739 0.4980580 0.8306681 0.4754171
#> [36,] 0.5820398 0.8783356 0.4411254 0.7083957 0.7295459 0.9258242 0.7494455
#> [37,] 0.5708684 0.8852519 0.5153235 0.7394159 0.8150584 0.9328585 0.7628806
#> [38,] 0.6532613 0.9256218 0.5226015 0.8179568 0.8570993 0.9595339 0.8373401
#> [39,] 0.7187087 0.9089421 0.5264049 0.8288120 0.8521170 0.9549295 0.8288069
#> [40,] 0.4613234 0.8255912 0.4002929 0.6232715 0.6660879 0.8924574 0.6237173
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6315667 0.4860008 0.7068970
#>  [2,] 0.6101292 0.3099930 0.5796289
#>  [3,] 0.5083362 0.3223110 0.5734574
#>  [4,] 0.6672527 0.3964039 0.7369208
#>  [5,] 0.6515968 0.4211880 0.7535561
#>  [6,] 0.5015578 0.3370009 0.6409463
#>  [7,] 0.6895430 0.3833207 0.6837226
#>  [8,] 0.3600845 0.2723192 0.5036641
#>  [9,] 0.6230173 0.4655258 0.7362497
#> [10,] 0.7319564 0.4759446 0.7950630
#> [11,] 0.6332018 0.5180768 0.7798570
#> [12,] 0.7048259 0.4103485 0.6092825
#> [13,] 0.6588046 0.3949058 0.7022033
#> [14,] 0.5062468 0.3777146 0.6900748
#> [15,] 0.6550193 0.4498182 0.7710584
#> [16,] 0.4710843 0.3008659 0.5249258
#> [17,] 0.6651975 0.3723939 0.5705251
#> [18,] 0.6593024 0.4308504 0.7174232
#> [19,] 0.7027050 0.5617236 0.8262927
#> [20,] 0.6620118 0.4200388 0.6911737
#> [21,] 0.6408785 0.5333697 0.7756912
#> [22,] 0.6215401 0.3785916 0.6511592
#> [23,] 0.4281940 0.2902379 0.5840842
#> [24,] 0.7042777 0.5080446 0.7756379
#> [25,] 0.4531811 0.3520785 0.6906183
#> [26,] 0.6605587 0.4861482 0.7729136
#> [27,] 0.6825043 0.3888308 0.6795542
#> [28,] 0.4409212 0.3032570 0.5134028
#> [29,] 0.6542938 0.5128211 0.7971565
#> [30,] 0.7091608 0.5542235 0.8323540
#> [31,] 0.5887589 0.3830326 0.7128591
#> [32,] 0.7083667 0.4720053 0.7032022
#> [33,] 0.4949179 0.4124189 0.6072831
#> [34,] 0.7013405 0.4741598 0.7039669
#> [35,] 0.4398527 0.2832700 0.5453473
#> [36,] 0.6549304 0.3968531 0.7020630
#> [37,] 0.6693547 0.4855353 0.7778758
#> [38,] 0.7189595 0.4875930 0.8099016
#> [39,] 0.7625058 0.5364422 0.7932159
#> [40,] 0.5568849 0.3815224 0.6584617
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4159195 0.5840805
#>  [2,] 0.4404737 0.5595263
#>  [3,] 0.4387417 0.5612583
#>  [4,] 0.4217815 0.5782185
#>  [5,] 0.4189395 0.5810605
#>  [6,] 0.4407516 0.5592484
#>  [7,] 0.4261539 0.5738461
#>  [8,] 0.4481407 0.5518593
#>  [9,] 0.4189065 0.5810935
#> [10,] 0.4116519 0.5883481
#> [11,] 0.4131003 0.5868997
#> [12,] 0.4261626 0.5738374
#> [13,] 0.4217571 0.5782429
#> [14,] 0.4325042 0.5674958
#> [15,] 0.4170452 0.5829548
#> [16,] 0.4435475 0.5564525
#> [17,] 0.4315121 0.5684879
#> [18,] 0.4212446 0.5787554
#> [19,] 0.4020461 0.5979539
#> [20,] 0.4215700 0.5784300
#> [21,] 0.4085006 0.5914994
#> [22,] 0.4238495 0.5761505
#> [23,] 0.4493860 0.5506140
#> [24,] 0.4130387 0.5869613
#> [25,] 0.4344552 0.5655448
#> [26,] 0.4173718 0.5826282
#> [27,] 0.4242611 0.5757389
#> [28,] 0.4422697 0.5577303
#> [29,] 0.4087154 0.5912846
#> [30,] 0.4029158 0.5970842
#> [31,] 0.4250897 0.5749103
#> [32,] 0.4176122 0.5823878
#> [33,] 0.4268499 0.5731501
#> [34,] 0.4139412 0.5860588
#> [35,] 0.4453942 0.5546058
#> [36,] 0.4220599 0.5779401
#> [37,] 0.4142117 0.5857883
#> [38,] 0.4102270 0.5897730
#> [39,] 0.4055459 0.5944541
#> [40,] 0.4316372 0.5683628
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]        [,3]        [,4]         [,5]      [,6]
#>  [1,]  0.50124695 1.8385391 -0.12299564  0.92614047  0.989391731 2.4047517
#>  [2,]  0.14803060 1.5435579 -0.87308158  0.75620457  0.486480269 2.2747107
#>  [3,] -0.22389102 1.2244780 -0.70925269  0.12383961  0.194708653 1.6562418
#>  [4,]  0.30623864 1.9739689 -0.26779033  0.87315858  1.027830975 2.6472171
#>  [5,]  0.31595191 2.2768647 -0.13875565  0.96126667  1.288249394 2.8762748
#>  [6,] -0.32986311 1.4389251 -0.62386172  0.53865326  0.676528137 2.0311957
#>  [7,]  0.54205704 1.9251348 -0.46408112  1.24722613  0.904539194 2.6398610
#>  [8,] -0.81115035 0.7785578 -0.76880968 -0.35813664 -0.365166500 0.9728567
#>  [9,]  0.22843960 1.9250110 -0.06260117  0.96233315  1.325199326 2.4924287
#> [10,]  0.70964392 2.5065861  0.01444651  1.59103157  1.740357827 3.2071387
#> [11,]  0.29118297 2.0211840  0.12253369  1.12004366  1.516176130 2.5442337
#> [12,]  0.67414655 1.4341695 -0.61594671  0.91134674  0.578611388 2.0925805
#> [13,]  0.33200756 1.9904144 -0.21989405  0.89453860  0.974495947 2.5157569
#> [14,] -0.40566431 1.6579546 -0.33515348  0.37719797  0.834487760 2.0856938
#> [15,]  0.24556652 2.0846757 -0.02822169  0.97979345  1.319870909 2.6931001
#> [16,] -0.31862897 0.9935487 -0.76413460  0.10930808 -0.066337969 1.5015243
#> [17,]  0.48616245 1.3326301 -0.73769803  0.72490430  0.322054987 1.9731828
#> [18,]  0.44191506 1.8288475 -0.25226152  1.06287508  1.187873068 2.5119009
#> [19,]  0.59443619 2.4314322  0.42142209  1.35110428  1.912452897 3.0295944
#> [20,]  0.46338625 1.8419868 -0.28730799  0.97656656  0.967089969 2.4221077
#> [21,]  0.38353403 2.0455923  0.27673926  1.09299586  1.459552542 2.4331577
#> [22,]  0.24976857 1.6333043 -0.29194245  0.59359703  0.481508573 2.0149209
#> [23,] -0.67720816 0.9530025 -0.87150168  0.09649124  0.160385458 1.4584347
#> [24,]  0.51080941 2.1268185 -0.04156853  1.22505885  1.494161583 2.6894737
#> [25,] -0.53811391 1.5724461 -0.34001657  0.23209131  0.709186985 1.9698808
#> [26,]  0.30133741 2.0272988 -0.08367056  1.13197370  1.468640652 2.7076255
#> [27,]  0.49629108 1.8724895 -0.40241338  1.02871574  0.848675162 2.5177910
#> [28,] -0.25322247 0.8423475 -0.70397095  0.04260571 -0.116138793 1.3775957
#> [29,]  0.36384551 2.2122475  0.26111056  1.07188582  1.606522278 2.7601221
#> [30,]  0.57624422 2.4895558  0.39002333  1.38076703  1.975494613 3.1269116
#> [31,] -0.04047804 1.7881670 -0.19336257  0.58556147  0.926950792 2.2683567
#> [32,]  0.64689178 1.9552128 -0.19849078  1.32031140  1.184692181 2.5302390
#> [33,] -0.18534293 1.1016372 -0.25655303  0.02227222  0.330075019 1.4205401
#> [34,]  0.69217111 1.7084605 -0.07995269  1.08080867  1.117212446 2.3052983
#> [35,] -0.47248111 1.0667119 -0.75742010  0.06872270 -0.007768094 1.5903698
#> [36,]  0.33115267 1.9767626 -0.23659575  0.88760517  0.992320003 2.5242459
#> [37,]  0.28539500 2.0431331  0.06131322  1.04293504  1.483219754 2.6314513
#> [38,]  0.63340583 2.5213028  0.09046747  1.50256571  1.791403445 3.1659839
#> [39,]  0.93806514 2.3007864  0.10571778  1.57723066  1.751302550 3.0534096
#> [40,] -0.15501600 1.5546975 -0.40424464  0.50345756  0.690543740 2.1160915
#>              [,7]         [,8]        [,9]      [,10]
#>  [1,]  0.94994767  0.538943971 -0.05601144 0.88036101
#>  [2,]  0.50731639  0.447855233 -0.80015202 0.32125024
#>  [3,]  0.09075443  0.033347929 -0.74317140 0.29597158
#>  [4,]  0.84990482  0.695785497 -0.42047178 1.03002512
#>  [5,]  1.12053059  0.626065704 -0.31789838 1.11766932
#>  [6,]  0.53936488  0.006231047 -0.67668801 0.57947372
#>  [7,]  1.08598966  0.797983469 -0.47547712 0.77093252
#>  [8,] -0.29104428 -0.574997533 -0.98288737 0.01465672
#>  [9,]  1.08888262  0.502375109 -0.13811607 1.02656657
#> [10,]  1.61041187  1.004571146 -0.09629592 1.35571867
#> [11,]  1.26347305  0.545977379  0.07233864 1.26483350
#> [12,]  0.51139712  0.870385322 -0.36252502 0.44429723
#> [13,]  1.04735308  0.657971502 -0.42673702 0.85781180
#> [14,]  0.60484757  0.024988456 -0.49925999 0.80046924
#> [15,]  1.10206730  0.641176247 -0.20140530 1.21429702
#> [16,] -0.14767359 -0.115791959 -0.84317788 0.09978585
#> [17,]  0.32972358  0.686543029 -0.52196043 0.28399407
#> [18,]  0.93363650  0.660187248 -0.27838233 0.93171547
#> [19,]  1.55501316  0.860212374  0.24816022 1.55957730
#> [20,]  0.87186556  0.672272342 -0.32261398 0.80561202
#> [21,]  1.28227491  0.579179277  0.13367769 1.24073108
#> [22,]  0.50359277  0.496090446 -0.49553048 0.62413846
#> [23,]  0.01973503 -0.289223278 -0.89422877 0.33956239
#> [24,]  1.31208047  0.867751943  0.03218110 1.24042442
#> [25,]  0.53271344 -0.187825999 -0.60991557 0.80301162
#> [26,]  1.22493073  0.665784999 -0.05542121 1.22483667
#> [27,]  0.93844285  0.765304639 -0.45222954 0.75172376
#> [28,] -0.19269019 -0.237424109 -0.83183609 0.05362422
#> [29,]  1.27381628  0.637967094  0.05129555 1.36861643
#> [30,]  1.58934262  0.891311933  0.21775035 1.60240332
#> [31,]  0.80223536  0.358836989 -0.47669607 0.90931091
#> [32,]  1.16778245  0.887464890 -0.11209601 0.86259343
#> [33,]  0.06624253 -0.020328982 -0.35397450 0.43590611
#> [34,]  0.89554952  0.853689435 -0.10345285 0.86626008
#> [35,] -0.09841099 -0.241759830 -0.92829884 0.18188881
#> [36,]  1.09565712  0.640782852 -0.41859472 0.85714098
#> [37,]  1.16853744  0.705268113 -0.05787505 1.25333037
#> [38,]  1.63856880  0.939306020 -0.04963800 1.44937112
#> [39,]  1.57719525  1.166466915  0.14602778 1.34442024
#> [40,]  0.50535654  0.228529198 -0.48309126 0.65644636
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.04657805 0.2929691
#>  [2,] -0.01620930 0.2230307
#>  [3,] -0.03192072 0.2143497
#>  [4,] -0.03999039 0.2754742
#>  [5,] -0.03647342 0.2906550
#>  [6,] -0.01127955 0.2268328
#>  [7,] -0.02806273 0.2694980
#>  [8,] -0.02893273 0.1792533
#>  [9,] -0.03635475 0.2909091
#> [10,] -0.04092027 0.3162203
#> [11,] -0.04222578 0.3089379
#> [12,] -0.04461128 0.2529137
#> [13,] -0.03637508 0.2791895
#> [14,] -0.02311825 0.2485229
#> [15,] -0.04065942 0.2942557
#> [16,] -0.03113235 0.1956446
#> [17,] -0.04064209 0.2350425
#> [18,] -0.03764831 0.2800181
#> [19,] -0.05757495 0.3393720
#> [20,] -0.03972164 0.2766100
#> [21,] -0.05116841 0.3189989
#> [22,] -0.04783816 0.2591523
#> [23,] -0.01425446 0.1888975
#> [24,] -0.04355288 0.3078649
#> [25,] -0.02183025 0.2418663
#> [26,] -0.03627468 0.2972969
#> [27,] -0.03516979 0.2701353
#> [28,] -0.03606357 0.1958920
#> [29,] -0.05060025 0.3186782
#> [30,] -0.05536569 0.3379650
#> [31,] -0.03472802 0.2671857
#> [32,] -0.03968712 0.2928963
#> [33,] -0.05371227 0.2410029
#> [34,] -0.05414750 0.2935488
#> [35,] -0.02494372 0.1943540
#> [36,] -0.03481010 0.2795130
#> [37,] -0.04395311 0.3026282
#> [38,] -0.04267186 0.3203552
#> [39,] -0.05360103 0.3288083
#> [40,] -0.02949096 0.2456837
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4159195  0.4159195
#>  [2,] -0.4404737  0.4404737
#>  [3,] -0.4387417  0.4387417
#>  [4,] -0.4217815  0.4217815
#>  [5,] -0.4189395  0.4189395
#>  [6,]  0.5592484 -0.5592484
#>  [7,]  0.5738461 -0.5738461
#>  [8,] -0.4481407  0.4481407
#>  [9,]  0.5810935 -0.5810935
#> [10,]  0.5883481 -0.5883481
#> [11,]  0.5868997 -0.5868997
#> [12,]  0.5738374 -0.5738374
#> [13,] -0.4217571  0.4217571
#> [14,]  0.5674958 -0.5674958
#> [15,] -0.4170452  0.4170452
#> [16,]  0.5564525 -0.5564525
#> [17,]  0.5684879 -0.5684879
#> [18,] -0.4212446  0.4212446
#> [19,] -0.4020461  0.4020461
#> [20,]  0.5784300 -0.5784300
#> [21,] -0.4085006  0.4085006
#> [22,] -0.4238495  0.4238495
#> [23,]  0.5506140 -0.5506140
#> [24,] -0.4130387  0.4130387
#> [25,]  0.5655448 -0.5655448
#> [26,]  0.5826282 -0.5826282
#> [27,]  0.5757389 -0.5757389
#> [28,]  0.5577303 -0.5577303
#> [29,] -0.4087154  0.4087154
#> [30,] -0.4029158  0.4029158
#> [31,] -0.4250897  0.4250897
#> [32,] -0.4176122  0.4176122
#> [33,] -0.4268499  0.4268499
#> [34,]  0.5860588 -0.5860588
#> [35,]  0.5546058 -0.5546058
#> [36,] -0.4220599  0.4220599
#> [37,] -0.4142117  0.4142117
#> [38,]  0.5897730 -0.5897730
#> [39,] -0.4055459  0.4055459
#> [40,] -0.4316372  0.4316372
#> 
#> $L
#> [1] 0.6909654 0.7006580 0.7237448 0.6910955 0.6996759 0.6820929
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.6086957