Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13          V14
#>  [1,] -0.3281516 -0.04489060 -0.11789645  0.03050916  0.01754629  0.002421112
#>  [2,] -0.3225846 -0.13465737 -0.10878842 -0.09291883 -0.04314269 -0.098730068
#>  [3,] -0.2458758 -0.22774829 -0.19844639 -0.18097755 -0.24335635 -0.103201227
#>  [4,] -0.2771761 -0.05596549 -0.05332911 -0.19170679 -0.03365773 -0.120030869
#>  [5,] -0.2793324 -0.17716215 -0.14554088 -0.01544016 -0.18892002 -0.128025137
#>  [6,] -0.4073964 -0.19128077 -0.08052084 -0.04571471 -0.03101139  0.004321821
#>  [7,] -0.2672602 -0.08243495 -0.07376168 -0.07025127 -0.07085741 -0.085052536
#>  [8,] -0.2716474 -0.19786515 -0.20047943 -0.01462503 -0.04227586 -0.006813016
#>  [9,] -0.3446823 -0.09565284 -0.11638169 -0.02731614  0.00290985 -0.126946502
#> [10,] -0.2876974 -0.15662785 -0.25523021 -0.26899919 -0.21160121 -0.182625398
#>                 V15          V16           V17         V18          V19
#>  [1,]  9.629487e-03  0.091333549  0.0479552113  0.03924843  0.156897022
#>  [2,] -2.304460e-02 -0.051428335 -0.0926127022  0.04681962  0.101384525
#>  [3,] -5.370924e-02 -0.043471930 -0.0592764241  0.02424124 -0.077922980
#>  [4,]  1.952667e-02  0.012096296  0.0231442350  0.07194455  0.062979117
#>  [5,] -1.827533e-01 -0.030045843 -0.1390360398 -0.14718513 -0.098956030
#>  [6,] -2.383780e-02 -0.068388225  0.0988503926 -0.04552892  0.014667919
#>  [7,] -1.043519e-01 -0.053246616  0.1207911771  0.00391230  0.024961498
#>  [8,] -1.172657e-02 -0.144431034  0.0226701454  0.05382439 -0.042029117
#>  [9,]  2.318559e-06  0.009853058 -0.0465730990 -0.08581756 -0.022333570
#> [10,] -8.997873e-02 -0.128141045 -0.0009080614  0.01191680  0.006264536
#>               V2         V20          V21          V22        V23        V24
#>  [1,] -0.2548031  0.13557789  0.201008954  0.163307013 0.23509752 0.07533522
#>  [2,] -0.3815507  0.04440889  0.151200558  0.166764072 0.10936180 0.04037869
#>  [3,] -0.3888347 -0.04841259  0.099564255  0.046502602 0.06417656 0.12011692
#>  [4,] -0.3316345  0.09917636 -0.001223796  0.132749867 0.10541889 0.15872868
#>  [5,] -0.2146182  0.03652877  0.038435509  0.058968606 0.07263478 0.03851704
#>  [6,] -0.4544394  0.05947320  0.191765630  0.223849841 0.19001354 0.17032338
#>  [7,] -0.3058557  0.18436620  0.081725640  0.194745110 0.08453465 0.10448831
#>  [8,] -0.2524665  0.08912603  0.088122184  0.128125869 0.02297630 0.16907508
#>  [9,] -0.2710012 -0.04132324  0.059527241 -0.044945544 0.03053982 0.20549518
#> [10,] -0.1880252 -0.10662854 -0.123843526 -0.008754313 0.03845022 0.03672011
#>               V25         V26         V27         V28         V29         V3
#>  [1,]  0.16465171  0.11588521  0.15749987  0.16180278  0.18818653 -0.2500320
#>  [2,]  0.18845264  0.21934590  0.02114618  0.10929217  0.08288305 -0.2927877
#>  [3,]  0.12935228  0.11118848  0.20598606  0.13475039  0.04715769 -0.2552039
#>  [4,]  0.03963792  0.12876395 -0.01131610  0.10987813 -0.06392042 -0.2950364
#>  [5,]  0.12375748  0.13260581  0.11789819  0.10209373  0.03004666 -0.2756735
#>  [6,]  0.25373378  0.16959701  0.13565711  0.23175874  0.13003078 -0.3684607
#>  [7,]  0.20301471  0.11808883  0.25144340  0.15196782  0.15890004 -0.2676450
#>  [8,]  0.12091364 -0.03012624 -0.03068601 -0.05191513  0.01555204 -0.2102931
#>  [9,]  0.06815470  0.19371909  0.26108159  0.25256730  0.22289504 -0.3823107
#> [10,] -0.02339874  0.12351875  0.09368268  0.02307770  0.01211210 -0.2344955
#>               V30          V31         V32          V33          V34
#>  [1,]  0.12491986 -0.011480806 -0.02045964 -0.089546869 -0.026783336
#>  [2,]  0.02564710 -0.014467167 -0.09137769  0.029525685  0.003440948
#>  [3,] -0.04529171  0.056341254 -0.13558014 -0.015796268 -0.082830831
#>  [4,]  0.07019100 -0.008042216 -0.04668017 -0.151537079 -0.078915762
#>  [5,]  0.04247528  0.009050403 -0.11430184 -0.096922294 -0.090579012
#>  [6,]  0.12445220  0.138365251 -0.01144330  0.060959662  0.070278382
#>  [7,]  0.02704351  0.089313210  0.17423000  0.154977585  0.124966405
#>  [8,] -0.07104867  0.032997709 -0.02697081  0.006240056 -0.019844872
#>  [9,]  0.16372627  0.046886419  0.06918673 -0.048604459 -0.093632045
#> [10,]  0.06533815 -0.011433231  0.05320230  0.027315802  0.038277291
#>               V35         V36          V37         V38          V39         V4
#>  [1,]  0.05900325 -0.01095278  0.021857796  0.02525542 -0.049260188 -0.2346587
#>  [2,] -0.09058493 -0.06352435 -0.087766943 -0.11137106 -0.169723034 -0.2923940
#>  [3,] -0.18967075 -0.17879585 -0.141001470 -0.04405242 -0.054600334 -0.2386815
#>  [4,] -0.07161251 -0.02096817 -0.009318211  0.04180825  0.005294329 -0.2492413
#>  [5,] -0.12767808 -0.16369338 -0.121980522 -0.07693028 -0.104815040 -0.2103533
#>  [6,]  0.11011225  0.17817108  0.051312104  0.03881790 -0.017458930 -0.3691114
#>  [7,]  0.15022154  0.04348255  0.122020196 -0.01891604 -0.042562809 -0.2105764
#>  [8,] -0.06826742 -0.11586360 -0.116241050 -0.17569288 -0.068113515 -0.2916742
#>  [9,] -0.07817691 -0.01187735 -0.022736078 -0.12695095 -0.083384521 -0.3515721
#> [10,]  0.03420420  0.14152076  0.048855497 -0.07837586  0.023540399 -0.1872071
#>                 V40         V41          V42          V43           V44
#>  [1,]  0.0019613239 -0.09297262 -0.123687524 -0.150850249  0.0004312267
#>  [2,] -0.0896275033 -0.16701690 -0.036485585 -0.136650190 -0.1908494668
#>  [3,] -0.1638207746 -0.10524234 -0.122577791 -0.170344033 -0.0873067231
#>  [4,] -0.0942859657 -0.07950171 -0.026649442 -0.098921513 -0.1956667425
#>  [5,] -0.1037320830 -0.08364065 -0.183160830 -0.083672564 -0.0792892550
#>  [6,] -0.0084877724  0.02482780 -0.100439181 -0.161220385 -0.0749922204
#>  [7,]  0.0002157128 -0.01257692 -0.083992490  0.016569826 -0.1119315399
#>  [8,] -0.0549777946 -0.08509516 -0.010007793  0.009212151 -0.0841119053
#>  [9,] -0.2172068164 -0.18274979 -0.009350289  0.019810481 -0.0544397689
#> [10,] -0.1569944995 -0.05889916 -0.078297511 -0.111030373 -0.2349515627
#>               V45         V46        V47         V48        V49         V5
#>  [1,] -0.08586661 -0.01609718 -0.1841477 -0.17811876 -0.2297970 -0.2489495
#>  [2,] -0.17792068 -0.09226588 -0.2164478 -0.18389919 -0.3699585 -0.1956981
#>  [3,] -0.23550910 -0.14548935 -0.1679077 -0.34169470 -0.2099993 -0.1642275
#>  [4,] -0.17584476 -0.05501219 -0.1331607 -0.26229259 -0.2502579 -0.2054503
#>  [5,] -0.13356875 -0.12469133 -0.2828111 -0.29895183 -0.3126839 -0.2357700
#>  [6,] -0.07124975 -0.12768249 -0.1361869 -0.30700011 -0.3710216 -0.3464041
#>  [7,] -0.12876671 -0.16727855 -0.1777420 -0.21850462 -0.2306040 -0.1344973
#>  [8,] -0.14336985 -0.19099632 -0.2138649 -0.08296189 -0.2522314 -0.1808180
#>  [9,] -0.09436024 -0.21407255 -0.2265596 -0.11668252 -0.1755972 -0.2869255
#> [10,] -0.10779844 -0.13482263 -0.1149114 -0.27654652 -0.2183318 -0.2457063
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2413645 -0.3643473 -0.3170426 -0.3558894 -0.4100499 -0.3150764
#>  [2,] -0.3537714 -0.2505074 -0.3029830 -0.4194114 -0.3446252 -0.3714895
#>  [3,] -0.2422977 -0.2816438 -0.2829621 -0.3081106 -0.3325498 -0.3877205
#>  [4,] -0.2878378 -0.3855298 -0.3117158 -0.3620406 -0.2013223 -0.3361163
#>  [5,] -0.3113711 -0.3178744 -0.2858869 -0.4341827 -0.3679493 -0.3390636
#>  [6,] -0.4090968 -0.4267909 -0.3509739 -0.3399610 -0.4151280 -0.4717548
#>  [7,] -0.3116732 -0.3256905 -0.4133447 -0.2484435 -0.3127917 -0.3745998
#>  [8,] -0.3049710 -0.3104554 -0.3567581 -0.1881201 -0.3344515 -0.1738574
#>  [9,] -0.3805848 -0.2518532 -0.3015889 -0.3994617 -0.3552510 -0.4055697
#> [10,] -0.3957600 -0.3857433 -0.3126651 -0.3618571 -0.3210384 -0.3618184
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3314286 -0.3642372 -0.2461961 -0.4175700 -0.1863730 -0.3272171
#>  [2,] -0.3076481 -0.3049587 -0.4078829 -0.3043501 -0.1618674 -0.4104997
#>  [3,] -0.3377143 -0.3010299 -0.3919578 -0.2767051 -0.1748188 -0.3195427
#>  [4,] -0.2499853 -0.4016880 -0.2337608 -0.2596512 -0.2157163 -0.2218127
#>  [5,] -0.4552022 -0.4091465 -0.3454634 -0.3980811 -0.2503756 -0.4048656
#>  [6,] -0.4024446 -0.4166980 -0.3990670 -0.5120490 -0.1739749 -0.3738956
#>  [7,] -0.3952198 -0.3356100 -0.3687734 -0.3843057 -0.2051669 -0.3543763
#>  [8,] -0.2872415 -0.2315699 -0.2287791 -0.1977327 -0.2410152 -0.3320326
#>  [9,] -0.2965913 -0.2548235 -0.3081160 -0.2604452 -0.2883373 -0.3755972
#> [10,] -0.2791231 -0.2929953 -0.3659697 -0.3920556 -0.2622930 -0.2592662
#>               V7          V8          V9
#>  [1,] -0.2500881 -0.18839863 -0.16988038
#>  [2,] -0.2295070 -0.25082386 -0.23683526
#>  [3,] -0.2179539 -0.20356570 -0.22348725
#>  [4,] -0.2059706 -0.06705778 -0.13986098
#>  [5,] -0.1508627 -0.16674916 -0.15550765
#>  [6,] -0.1221293 -0.22764754 -0.08154310
#>  [7,] -0.1815359 -0.16410653 -0.06027008
#>  [8,] -0.1610606 -0.11340703 -0.21294213
#>  [9,] -0.2170039 -0.10054312 -0.09992992
#> [10,] -0.1490718 -0.25184580 -0.13542896
#> 
#> $W[[2]]
#>            [,1]        [,2]       [,3]         [,4]        [,5]        [,6]
#> [1,]  0.1533224 -0.03668313 0.01090422  0.006684534  0.10759622  0.05835563
#> [2,] -0.1257687 -0.01574248 0.02052731 -0.009710602 -0.08928275 -0.11792876
#>            [,7]        [,8]         [,9]       [,10]
#> [1,]  0.1403787  0.09151805  0.006746039 -0.15641437
#> [2,] -0.1449925 -0.07867393 -0.133125072  0.06370012
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -1.831200e-04 -0.0009364359 -0.0011241300 -0.0010542503 -0.0011600659
#>  [2,]  2.775943e-04  0.0015438759  0.0018674085  0.0017763609  0.0019414040
#>  [3,]  1.678411e-04  0.0010353631  0.0012458166  0.0012078975  0.0012825999
#>  [4,]  1.740981e-04  0.0010002886  0.0012062382  0.0011532827  0.0012425817
#>  [5,] -2.381187e-04 -0.0015147996 -0.0018455176 -0.0017789527 -0.0018329335
#>  [6,] -3.847287e-05 -0.0001827377 -0.0002278471 -0.0002129242 -0.0002335864
#>  [7,] -2.336201e-04 -0.0011865205 -0.0014610107 -0.0014045334 -0.0015291839
#>  [8,] -1.891859e-04 -0.0011326142 -0.0013811376 -0.0013211063 -0.0013988441
#>  [9,] -4.554993e-05 -0.0002448388 -0.0003041793 -0.0002773492 -0.0003010869
#> [10,]  4.626945e-04  0.0028029094  0.0033145469  0.0031667251  0.0033622890
#>                 V14           V15           V16           V17           V18
#>  [1,] -0.0010925360 -0.0007589454 -0.0005221003 -4.477718e-04 -0.0004896685
#>  [2,]  0.0017776331  0.0012220547  0.0008132077  6.912231e-04  0.0008533802
#>  [3,]  0.0011455163  0.0007953767  0.0005090308  4.306479e-04  0.0005762298
#>  [4,]  0.0011199545  0.0007701436  0.0005120928  4.330713e-04  0.0005499237
#>  [5,] -0.0016102362 -0.0010928804 -0.0006844634 -5.603747e-04 -0.0007051353
#>  [6,] -0.0002312153 -0.0001697513 -0.0001223552 -1.137983e-04 -0.0001078545
#>  [7,] -0.0014777550 -0.0010957320 -0.0008380214 -7.932210e-04 -0.0008818374
#>  [8,] -0.0012475799 -0.0008423898 -0.0005203695 -4.286850e-04 -0.0005361098
#>  [9,] -0.0002834790 -0.0001862252 -0.0001031201 -7.204681e-05 -0.0000665959
#> [10,]  0.0029649579  0.0019767005  0.0011455151  8.843674e-04  0.0012579964
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0010766981 -2.327458e-04 -0.0014375659 -0.0018506009 -0.0020576267
#>  [2,]  0.0018176352  3.481957e-04  0.0024041695  0.0030722325  0.0034478104
#>  [3,]  0.0011832524  2.099889e-04  0.0015200568  0.0019853847  0.0022872200
#>  [4,]  0.0011583575  2.170711e-04  0.0015239385  0.0019533522  0.0022008339
#>  [5,] -0.0016140402 -3.156970e-04 -0.0021804373 -0.0028060055 -0.0031311568
#>  [6,] -0.0002338218 -5.354268e-05 -0.0003276636 -0.0004116557 -0.0004309156
#>  [7,] -0.0016116006 -3.037479e-04 -0.0021314532 -0.0026579179 -0.0028561064
#>  [8,] -0.0012501305 -2.488165e-04 -0.0016807535 -0.0021659395 -0.0024160407
#>  [9,] -0.0002569797 -6.534891e-05 -0.0003886527 -0.0004871010 -0.0005107013
#> [10,]  0.0029296326  5.717972e-04  0.0038465783  0.0051486973  0.0060242454
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0019167977 -0.0017599550 -0.0014463234 -0.0016686290 -0.0018758716
#>  [2,]  0.0032775931  0.0030906610  0.0025677533  0.0028999719  0.0031870248
#>  [3,]  0.0022054816  0.0021124325  0.0017886330  0.0019836240  0.0021836702
#>  [4,]  0.0021037656  0.0020056984  0.0016792432  0.0018855851  0.0020736355
#>  [5,] -0.0028805930 -0.0026241373 -0.0021079827 -0.0023751665 -0.0026880769
#>  [6,] -0.0003873723 -0.0003365112 -0.0002634504 -0.0003022315 -0.0003374646
#>  [7,] -0.0026828931 -0.0025191259 -0.0021159015 -0.0023566932 -0.0025536448
#>  [8,] -0.0022314149 -0.0020232774 -0.0016020834 -0.0018063845 -0.0020331517
#>  [9,] -0.0004396631 -0.0003490563 -0.0002296351 -0.0002759058 -0.0003271536
#> [10,]  0.0058236111  0.0055710696  0.0047041511  0.0052688289  0.0058639573
#>                 V28           V29            V3           V30           V31
#>  [1,] -0.0017728311 -0.0012900988 -2.727314e-04 -8.566114e-04 -5.736287e-04
#>  [2,]  0.0029809053  0.0022480440  4.239967e-04  1.437950e-03  9.220875e-04
#>  [3,]  0.0020693406  0.0015930817  2.674151e-04  9.911869e-04  6.349222e-04
#>  [4,]  0.0019533139  0.0014896966  2.660887e-04  9.478862e-04  6.020210e-04
#>  [5,] -0.0025857701 -0.0018629447 -3.772818e-04 -1.005856e-03 -4.815662e-04
#>  [6,] -0.0003311107 -0.0002335115 -5.780193e-05 -1.521608e-04 -8.909747e-05
#>  [7,] -0.0024303664 -0.0018774553 -3.512730e-04 -1.274066e-03 -8.680635e-04
#>  [8,] -0.0019411224 -0.0013870386 -2.978399e-04 -7.533749e-04 -3.573229e-04
#>  [9,] -0.0003224054 -0.0001830948 -7.068111e-05 -5.771988e-05  3.205097e-05
#> [10,]  0.0055848219  0.0042684544  7.302074e-04  2.683740e-03  1.731852e-03
#>                 V32           V33           V34           V35           V36
#>  [1,] -6.325216e-04 -2.314109e-04  2.284579e-04  0.0005387955  0.0008822334
#>  [2,]  9.970060e-04  3.868998e-04 -3.233266e-04 -0.0008656731 -0.0013727403
#>  [3,]  6.446206e-04  2.271299e-04 -2.210061e-04 -0.0005606649 -0.0008597901
#>  [4,]  6.371909e-04  2.458960e-04 -2.031010e-04 -0.0005465647 -0.0008620209
#>  [5,] -5.894410e-04  8.922741e-05  7.997083e-04  0.0013245657  0.0017487096
#>  [6,] -1.088067e-04 -2.251933e-05  6.685442e-05  0.0001215978  0.0002002581
#>  [7,] -9.460797e-04 -5.004847e-04 -3.436828e-05  0.0003133465  0.0007548461
#>  [8,] -4.395444e-04  5.916596e-05  6.029580e-04  0.0009946485  0.0013515498
#>  [9,] -1.585357e-05  1.145577e-04  2.566468e-04  0.0003444023  0.0004338402
#> [10,]  1.762541e-03  6.006058e-04 -6.844280e-04 -0.0015857029 -0.0024498015
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0006243356 -3.016689e-04 -0.0007567738 -3.020841e-04 -2.862110e-04
#>  [2,] -0.0009593372  4.900825e-04  0.0012408629  4.558008e-04  5.633069e-04
#>  [3,] -0.0005729960  3.149403e-04  0.0008062980  2.823619e-04  4.162585e-04
#>  [4,] -0.0005952654  3.080891e-04  0.0007802718  2.842379e-04  3.662787e-04
#>  [5,]  0.0011750320 -2.670881e-04 -0.0009961441 -3.874693e-04 -3.738965e-04
#>  [6,]  0.0001414311 -6.321780e-05 -0.0001448820 -6.173514e-05 -3.432507e-05
#>  [7,]  0.0004989150 -5.569239e-04 -0.0010547624 -3.870536e-04 -4.885325e-04
#>  [8,]  0.0009441635 -1.987135e-04 -0.0007587887 -3.076421e-04 -2.601007e-04
#>  [9,]  0.0003167554 -4.757347e-06 -0.0001385319 -7.089241e-05  1.583013e-05
#> [10,] -0.0016870745  8.007959e-04  0.0021229458  7.740154e-04  1.020108e-03
#>                 V41           V42           V43           V44           V45
#>  [1,] -3.980481e-04 -0.0007776018 -0.0010559958 -0.0010032785 -0.0011481671
#>  [2,]  6.317050e-04  0.0012318222  0.0017108732  0.0016184203  0.0017995754
#>  [3,]  4.013961e-04  0.0007574240  0.0010892786  0.0010284645  0.0011087798
#>  [4,]  3.873003e-04  0.0007627766  0.0010757590  0.0010212149  0.0011287527
#>  [5,] -4.130530e-04 -0.0010143267 -0.0014699852 -0.0014576433 -0.0016361104
#>  [6,] -7.347951e-05 -0.0001714446 -0.0002182122 -0.0002060275 -0.0002392664
#>  [7,] -5.718902e-04 -0.0010538584 -0.0013727042 -0.0012721425 -0.0014089793
#>  [8,] -3.128988e-04 -0.0007875813 -0.0011468183 -0.0011161898 -0.0012531231
#>  [9,] -3.686281e-05 -0.0001810464 -0.0002578126 -0.0002621537 -0.0003139446
#> [10,]  1.063499e-03  0.0020330657  0.0029906389  0.0028259966  0.0030696678
#>                 V46           V47           V48           V49            V5
#>  [1,] -0.0010383952 -0.0006542552 -4.606295e-04 -3.076740e-04 -4.186894e-04
#>  [2,]  0.0015772670  0.0009858749  6.976777e-04  4.772439e-04  6.671427e-04
#>  [3,]  0.0009804945  0.0006282624  4.325157e-04  3.002341e-04  4.292617e-04
#>  [4,]  0.0009928717  0.0006238297  4.382934e-04  3.002141e-04  4.217661e-04
#>  [5,] -0.0014215160 -0.0009230072 -6.362809e-04 -4.395317e-04 -5.984046e-04
#>  [6,] -0.0002153251 -0.0001381734 -9.614054e-05 -6.396713e-05 -8.310631e-05
#>  [7,] -0.0012622144 -0.0008008551 -5.640569e-04 -3.774858e-04 -5.384353e-04
#>  [8,] -0.0010982825 -0.0007053346 -4.886195e-04 -3.392476e-04 -4.659127e-04
#>  [9,] -0.0002700576 -0.0001693894 -1.183422e-04 -8.258028e-05 -1.035574e-04
#> [10,]  0.0027456912  0.0017411411  1.192166e-03  8.254203e-04  1.170953e-03
#>                 V50           V51           V52           V53           V54
#>  [1,] -9.533901e-05 -1.045461e-04 -7.143524e-05 -4.070561e-05 -5.057036e-05
#>  [2,]  1.442775e-04  1.607013e-04  1.041012e-04  6.137615e-05  7.896471e-05
#>  [3,]  8.963810e-05  1.021498e-04  6.310773e-05  3.732216e-05  4.938610e-05
#>  [4,]  9.022228e-05  1.015130e-04  6.493614e-05  3.787608e-05  4.917723e-05
#>  [5,] -1.187516e-04 -1.477630e-04 -9.701749e-05 -5.288272e-05 -6.514120e-05
#>  [6,] -1.896015e-05 -2.311863e-05 -1.716434e-05 -9.831657e-06 -1.082471e-05
#>  [7,] -1.181272e-04 -1.339028e-04 -9.262533e-05 -5.617890e-05 -6.774279e-05
#>  [8,] -9.441524e-05 -1.156384e-04 -7.611741e-05 -4.230352e-05 -5.266069e-05
#>  [9,] -2.088442e-05 -2.800545e-05 -2.079341e-05 -1.103247e-05 -1.232251e-05
#> [10,]  2.482177e-04  2.791031e-04  1.733824e-04  1.003396e-04  1.349166e-04
#>                 V55           V56           V57           V58           V59
#>  [1,] -3.046119e-05 -2.404769e-05 -1.754103e-05 -4.088505e-05 -3.619624e-05
#>  [2,]  4.765258e-05  3.723808e-05  2.382770e-05  6.078591e-05  5.462182e-05
#>  [3,]  2.838147e-05  2.339238e-05  1.586334e-05  3.797578e-05  3.390282e-05
#>  [4,]  2.925861e-05  2.294709e-05  1.469623e-05  3.767464e-05  3.414153e-05
#>  [5,] -3.522145e-05 -3.003608e-05 -1.866517e-05 -5.307321e-05 -4.317622e-05
#>  [6,] -6.459705e-06 -4.836421e-06 -4.029182e-06 -9.269793e-06 -7.194851e-06
#>  [7,] -4.236354e-05 -3.160739e-05 -2.387983e-05 -5.352150e-05 -4.680869e-05
#>  [8,] -2.902379e-05 -2.388006e-05 -1.501959e-05 -4.261459e-05 -3.521059e-05
#>  [9,] -6.477316e-06 -5.170232e-06 -3.135982e-06 -1.076494e-05 -7.573396e-06
#> [10,]  7.506151e-05  6.265009e-05  4.344946e-05  1.037049e-04  9.379933e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -3.097954e-04 -2.831876e-05 -3.178704e-04 -4.646954e-04 -0.0008102859
#>  [2,]  5.181604e-04  4.126106e-05  5.296057e-04  7.736674e-04  0.0013045878
#>  [3,]  3.546467e-04  2.578297e-05  3.638704e-04  5.331338e-04  0.0008598076
#>  [4,]  3.340999e-04  2.598068e-05  3.433628e-04  5.012402e-04  0.0008386060
#>  [5,] -4.979249e-04 -3.152459e-05 -4.688987e-04 -7.608970e-04 -0.0012847043
#>  [6,] -6.343440e-05 -5.062693e-06 -5.913885e-05 -9.697581e-05 -0.0001669744
#>  [7,] -4.364672e-04 -3.512664e-05 -4.393763e-04 -6.248937e-04 -0.0010291516
#>  [8,] -3.707874e-04 -2.535779e-05 -3.536839e-04 -5.662940e-04 -0.0009624351
#>  [9,] -7.565922e-05 -4.773976e-06 -6.084231e-05 -1.215475e-04 -0.0002268022
#> [10,]  9.251290e-04  7.350688e-05  9.521664e-04  1.411128e-03  0.0023422092
#> 
#> $vW[[2]]
#>            [,1]       [,2]        [,3]        [,4]        [,5]      [,6]
#> [1,] -0.0785615 -0.0569041 -0.04100808 -0.04964032 -0.03947536 -0.077267
#> [2,]  0.0785615  0.0569041  0.04100808  0.04964032  0.03947536  0.077267
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.07334506 -0.04072147 -0.06113718 -0.02092633
#> [2,]  0.07334506  0.04072147  0.06113718  0.02092633
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5707021 0.4435990 0.2317126 0.2324321 0.3042559 0.6901541 0.5837995
#>  [8] 0.1925860 0.3425146 0.2933127
#> 
#> $B[[2]]
#> [1]  0.20806311 -0.04595733
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -0.0020171033  0.0033865912  0.0022835400  0.0021818815 -0.0026846857
#>  [6] -0.0003860862 -0.0029168174 -0.0020695298 -0.0003222405  0.0060352341
#> 
#> $vB[[2]]
#> [1] -0.1015295  0.1015295
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#>  [2,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#>  [3,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>  [4,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#>  [5,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#>  [6,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [7,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#>  [8,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#>  [9,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [10,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [11,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [12,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [13,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [14,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [15,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [16,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [17,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [18,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [19,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [20,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [21,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [22,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [23,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [24,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [25,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [26,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [27,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [28,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [29,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [30,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [31,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [32,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [33,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [34,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [35,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [36,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [37,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [38,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [39,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [40,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#>  [2,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#>  [3,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>  [4,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#>  [5,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#>  [6,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [7,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#>  [8,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#>  [9,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [10,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [11,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [12,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [13,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [14,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [15,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [16,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [17,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [18,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [19,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [20,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [21,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [22,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [23,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [24,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [25,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [26,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [27,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [28,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [29,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [30,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [31,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [32,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [33,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [34,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [35,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [36,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [37,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [38,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [39,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [40,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#>  [2,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#>  [3,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>  [4,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#>  [5,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#>  [6,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [7,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#>  [8,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#>  [9,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [10,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [11,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [12,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [13,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [14,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [15,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [16,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [17,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [18,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [19,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [20,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [21,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [22,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [23,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [24,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [25,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [26,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [27,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [28,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [29,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [30,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [31,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [32,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [33,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [34,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [35,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [36,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [37,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [38,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [39,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [40,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#>  [2,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#>  [3,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>  [4,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#>  [5,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#>  [6,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [7,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#>  [8,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#>  [9,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [10,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [11,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [12,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [13,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [14,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [15,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [16,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [17,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [18,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [19,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [20,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [21,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [22,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [23,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [24,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [25,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [26,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [27,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [28,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [29,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [30,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [31,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [32,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [33,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [34,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [35,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [36,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [37,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [38,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [39,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [40,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#>  [2,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#>  [3,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>  [4,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#>  [5,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#>  [6,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [7,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#>  [8,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#>  [9,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [10,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [11,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [12,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [13,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [14,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [15,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [16,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [17,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [18,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [19,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [20,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [21,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [22,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [23,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [24,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [25,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [26,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [27,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [28,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [29,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [30,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [31,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [32,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [33,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [34,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [35,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [36,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [37,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [38,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [39,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [40,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#>  [2,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#>  [3,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#>  [4,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#>  [5,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#>  [6,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [7,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#>  [8,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#>  [9,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [10,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [11,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [12,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [13,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [14,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [15,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [16,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [17,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [18,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [19,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [20,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [21,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [22,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [23,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [24,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [25,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [26,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [27,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [28,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [29,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [30,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [31,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [32,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [33,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [34,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [35,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [36,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [37,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [38,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [39,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [40,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8683054 0.7089852 0.5457490 0.6197368 0.4877823 0.8910147 0.8569186
#>  [2,] 0.7743683 0.5195325 0.3498667 0.4768985 0.3184168 0.8194941 0.8034625
#>  [3,] 0.8154632 0.5936097 0.3972482 0.5074956 0.4000911 0.8443413 0.8128890
#>  [4,] 0.8658443 0.6620298 0.4633346 0.6056473 0.4110647 0.8653215 0.8245935
#>  [5,] 0.7548187 0.5207545 0.3296528 0.4776122 0.3078549 0.8368438 0.8245206
#>  [6,] 0.8418008 0.7045300 0.5687602 0.6265129 0.5507541 0.8844002 0.8537928
#>  [7,] 0.7926320 0.5500395 0.3395255 0.5226624 0.2967304 0.7769417 0.7516003
#>  [8,] 0.7590303 0.5776840 0.4294224 0.5167959 0.4164642 0.7692733 0.7330478
#>  [9,] 0.8186337 0.6386106 0.4216615 0.5982398 0.4014548 0.8463658 0.8182876
#> [10,] 0.8232150 0.5876423 0.3871500 0.5683720 0.3276811 0.8199762 0.7621611
#> [11,] 0.8531051 0.6707585 0.4885810 0.6152838 0.4252780 0.8632604 0.8141317
#> [12,] 0.7689896 0.5128866 0.3360692 0.5335750 0.2886954 0.7692173 0.7328568
#> [13,] 0.8004996 0.5910707 0.3928640 0.5619994 0.3422916 0.8193139 0.7939181
#> [14,] 0.7972505 0.6499475 0.5301487 0.5596780 0.4999687 0.8515649 0.8268059
#> [15,] 0.7701998 0.4379070 0.2440916 0.4677133 0.2411644 0.7883582 0.7593056
#> [16,] 0.8345477 0.6265990 0.4029881 0.5702280 0.3464725 0.8466649 0.8168299
#> [17,] 0.8491054 0.7028357 0.5507055 0.5908471 0.5235626 0.8754853 0.8424264
#> [18,] 0.8084411 0.6619899 0.5456830 0.5720975 0.5104814 0.8597948 0.8317691
#> [19,] 0.7188085 0.4382841 0.2963007 0.4097917 0.2989841 0.8282397 0.8183894
#> [20,] 0.8321032 0.6818337 0.5645068 0.6085741 0.5232324 0.8789475 0.8484027
#> [21,] 0.8501833 0.6982962 0.5472384 0.6038511 0.5224007 0.8685173 0.8327585
#> [22,] 0.8136737 0.5858181 0.3458166 0.5302106 0.3050670 0.8521989 0.8453061
#> [23,] 0.7972433 0.6341038 0.4789754 0.5646435 0.4543869 0.8299716 0.7864804
#> [24,] 0.8220457 0.6057564 0.4260596 0.5484761 0.4098378 0.8592304 0.8188499
#> [25,] 0.6722300 0.4337748 0.2950819 0.4246413 0.2767164 0.7683610 0.7498688
#> [26,] 0.8320747 0.6469272 0.4714018 0.5936487 0.4285210 0.8514544 0.8114639
#> [27,] 0.8283952 0.6116370 0.3964503 0.5816546 0.3512013 0.8187750 0.7718475
#> [28,] 0.8528822 0.6974465 0.5244004 0.6004768 0.5004368 0.8764770 0.8418699
#> [29,] 0.8209580 0.5864420 0.3881274 0.5101946 0.3853611 0.8506831 0.8223885
#> [30,] 0.7366579 0.5128127 0.3180703 0.4931404 0.2832743 0.7192024 0.7159356
#> [31,] 0.8573289 0.7127418 0.5654943 0.6200908 0.5297963 0.9196886 0.8993936
#> [32,] 0.7957726 0.6372721 0.5160300 0.5575194 0.4890702 0.8472615 0.8279413
#> [33,] 0.8392203 0.5776784 0.3677979 0.5390066 0.3244413 0.9070433 0.8886299
#> [34,] 0.8424469 0.6787698 0.5071964 0.6054154 0.4773345 0.8794250 0.8448776
#> [35,] 0.7472298 0.4816472 0.3036187 0.4531967 0.2686245 0.7765309 0.7787815
#> [36,] 0.8671445 0.7293066 0.5743856 0.6550208 0.4927324 0.8857357 0.8542144
#> [37,] 0.8054877 0.6432165 0.4358433 0.5360777 0.4224726 0.8766102 0.8574195
#> [38,] 0.8637151 0.7238547 0.5819614 0.6524791 0.5003337 0.8769950 0.8422881
#> [39,] 0.8262911 0.6304453 0.4427132 0.5515553 0.4322163 0.8643990 0.8304542
#> [40,] 0.8561968 0.7161105 0.5706082 0.6117783 0.5415102 0.8824461 0.8524368
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5055790 0.7308077 0.4497741
#>  [2,] 0.3679174 0.6058411 0.3760082
#>  [3,] 0.4236135 0.6506166 0.3394881
#>  [4,] 0.4810676 0.6484060 0.4068002
#>  [5,] 0.4014675 0.5652631 0.4296465
#>  [6,] 0.5094402 0.7622014 0.5209528
#>  [7,] 0.4233257 0.5182719 0.2951736
#>  [8,] 0.4464922 0.5966044 0.3682320
#>  [9,] 0.4801870 0.5710253 0.4421025
#> [10,] 0.4378903 0.5598994 0.3050364
#> [11,] 0.4928976 0.6370600 0.3855793
#> [12,] 0.3905762 0.5417522 0.3663994
#> [13,] 0.4468363 0.5206605 0.3738943
#> [14,] 0.4673118 0.7317075 0.4991408
#> [15,] 0.3128702 0.4954546 0.2586075
#> [16,] 0.4501201 0.6085893 0.3745643
#> [17,] 0.5018837 0.7481727 0.4512815
#> [18,] 0.4728144 0.7423954 0.4815917
#> [19,] 0.3133571 0.6111453 0.4350386
#> [20,] 0.5046438 0.7451396 0.5253019
#> [21,] 0.4994291 0.7358736 0.4365396
#> [22,] 0.4320452 0.5166931 0.3949983
#> [23,] 0.4750463 0.5933924 0.4370641
#> [24,] 0.4392854 0.6177472 0.3744439
#> [25,] 0.3414039 0.4794523 0.4070724
#> [26,] 0.4680488 0.6613603 0.3938548
#> [27,] 0.4702728 0.5680908 0.3321362
#> [28,] 0.4922309 0.7267102 0.4382177
#> [29,] 0.4083432 0.6642667 0.3377937
#> [30,] 0.4272642 0.4460218 0.3457289
#> [31,] 0.5052206 0.7588716 0.5931763
#> [32,] 0.4526737 0.7213911 0.4675155
#> [33,] 0.3700850 0.6558350 0.5160336
#> [34,] 0.4772288 0.7403248 0.5054966
#> [35,] 0.3504756 0.5096389 0.3204010
#> [36,] 0.5580651 0.6849549 0.4784435
#> [37,] 0.4775407 0.6528525 0.4631886
#> [38,] 0.5554999 0.6872940 0.4611864
#> [39,] 0.4474358 0.6760037 0.3927078
#> [40,] 0.5180207 0.7504307 0.4764941
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5238813 0.4761187
#>  [2,] 0.5288242 0.4711758
#>  [3,] 0.5325426 0.4674574
#>  [4,] 0.5259488 0.4740512
#>  [5,] 0.5262093 0.4737907
#>  [6,] 0.5198377 0.4801623
#>  [7,] 0.5311031 0.4688969
#>  [8,] 0.5261446 0.4738554
#>  [9,] 0.5231362 0.4768638
#> [10,] 0.5301074 0.4698926
#> [11,] 0.5261352 0.4738648
#> [12,] 0.5252887 0.4747113
#> [13,] 0.5263569 0.4736431
#> [14,] 0.5203349 0.4796651
#> [15,] 0.5360339 0.4639661
#> [16,] 0.5272753 0.4727247
#> [17,] 0.5243507 0.4756493
#> [18,] 0.5215004 0.4784996
#> [19,] 0.5274282 0.4725718
#> [20,] 0.5192824 0.4807176
#> [21,] 0.5247949 0.4752051
#> [22,] 0.5278030 0.4721970
#> [23,] 0.5227028 0.4772972
#> [24,] 0.5293260 0.4706740
#> [25,] 0.5244050 0.4755950
#> [26,] 0.5260913 0.4739087
#> [27,] 0.5289752 0.4710248
#> [28,] 0.5249004 0.4750996
#> [29,] 0.5328650 0.4671350
#> [30,] 0.5262811 0.4737189
#> [31,] 0.5165635 0.4834365
#> [32,] 0.5224768 0.4775232
#> [33,] 0.5218671 0.4781329
#> [34,] 0.5202727 0.4797273
#> [35,] 0.5310898 0.4689102
#> [36,] 0.5208888 0.4791112
#> [37,] 0.5246012 0.4753988
#> [38,] 0.5216922 0.4783078
#> [39,] 0.5285123 0.4714877
#> [40,] 0.5228641 0.4771359
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]        [,2]        [,3]        [,4]          [,5]      [,6]
#>  [1,] 1.8860578  0.89046073  0.18350929  0.48843133 -0.0488805856 2.1011476
#>  [2,] 1.2331433  0.07816963 -0.61962514 -0.09247174 -0.7610570235 1.5129238
#>  [3,] 1.4859075  0.37890790 -0.41694413  0.02998469 -0.4050857226 1.6908907
#>  [4,] 1.8647038  0.67235270 -0.14692544  0.42905229 -0.3595658802 1.8602109
#>  [5,] 1.1244797  0.08306567 -0.70975572 -0.08961101 -0.8101671424 1.6349294
#>  [6,] 1.6716883  0.86896350  0.27679459  0.51728599  0.2037179244 2.0347754
#>  [7,] 1.3408643  0.20083021 -0.66540943  0.09071186 -0.8629163460 1.2479320
#>  [8,] 1.1473706  0.31327296 -0.28420829  0.06720873 -0.3373051123 1.2042124
#>  [9,] 1.5071177  0.56933876 -0.31595652  0.39813642 -0.3994071253 1.7063774
#> [10,] 1.5382833  0.35422712 -0.45930811  0.27521189 -0.7186920707 1.5161864
#> [11,] 1.7591653  0.71161772 -0.04568392  0.46957784 -0.3011433688 1.8426383
#> [12,] 1.2026145  0.05155791 -0.68086102  0.13450240 -0.9017288259 1.2038971
#> [13,] 1.3894195  0.36839333 -0.43528929  0.24928049 -0.6530985796 1.5117059
#> [14,] 1.3691976  0.61880833  0.12074107  0.23985525 -0.0001253989 1.7469281
#> [15,] 1.2094398 -0.24966094 -1.13037675 -0.12932657 -1.1463061235 1.3150576
#> [16,] 1.6182073  0.51765373 -0.39302991  0.28278135 -0.6345809919 1.7086791
#> [17,] 1.7276018  0.86083789  0.20352149  0.36746838  0.0943202910 1.9503543
#> [18,] 1.4399126  0.67217465  0.18324321  0.29041403  0.0419317519 1.8135868
#> [19,] 0.9385589 -0.24812895 -0.86497645 -0.36482666 -0.8521400574 1.5732028
#> [20,] 1.6006069  0.76221157  0.25947323  0.44132240  0.0929965749 1.9825006
#> [21,] 1.7360391  0.83919743  0.18951886  0.42153749  0.0896629332 1.8879126
#> [22,] 1.4740600  0.34670414 -0.63747923  0.12098964 -0.8232840040 1.7519527
#> [23,] 1.3691535  0.54986279 -0.08414821  0.26002916 -0.1829611712 1.5854261
#> [24,] 1.5302692  0.42950890 -0.29794642  0.19451531 -0.3646357384 1.8089128
#> [25,] 0.7182882 -0.26646625 -0.87082871 -0.30374890 -0.9608083596 1.1990794
#> [26,] 1.6004030  0.60555964 -0.11451768  0.37906968 -0.2878878485 1.7460543
#> [27,] 1.5742958  0.45419853 -0.42027767  0.32956958 -0.6137629458 1.5080700
#> [28,] 1.7573875  0.83516755  0.09767925  0.40745228  0.0017471815 1.9594834
#> [29,] 1.5228513  0.34927593 -0.45519058  0.04078399 -0.4668541861 1.7399690
#> [30,] 1.0286696  0.05126214 -0.76265396 -0.02744028 -0.9282773991 0.9405088
#> [31,] 1.7932796  0.90873760  0.26349111  0.48993370  0.1193267603 2.4381237
#> [32,] 1.3600796  0.56354365  0.06414200  0.23110053 -0.0437262282 1.7132822
#> [33,] 1.6524380  0.31325035 -0.54167555  0.15634428 -0.7334356386 2.2780562
#> [34,] 1.6765478  0.74812397  0.02878739  0.42808115 -0.0907240029 1.9869963
#> [35,] 1.0838921 -0.07344420 -0.83012481 -0.18776307 -1.0016122889 1.2455636
#> [36,] 1.8759435  0.99110741  0.29976723  0.64118273 -0.0290723680 2.0479047
#> [37,] 1.4209523  0.58935243 -0.25804939  0.14456212 -0.3126313284 1.9607142
#> [38,] 1.8464952  0.96366353  0.33083024  0.62995446  0.0013346638 1.9642765
#> [39,] 1.5595663  0.53412746 -0.23015796  0.20695678 -0.2728142108 1.8523173
#> [40,] 1.7840544  0.92524963  0.28433319  0.45479334  0.1664236952 2.0158010
#>            [,7]         [,8]        [,9]        [,10]
#>  [1,] 1.7899293  0.022316877  0.99872442 -0.201583486
#>  [2,] 1.4080770 -0.541161547  0.42986367 -0.506526319
#>  [3,] 1.4688922 -0.307956966  0.62175064 -0.665576255
#>  [4,] 1.5477842 -0.075765953  0.61203984 -0.377209037
#>  [5,] 1.5472805 -0.399354233  0.26255041 -0.283293722
#>  [6,] 1.7646641  0.037765233  1.16478684  0.083860496
#>  [7,] 1.1071657 -0.309135749  0.07312020 -0.870387636
#>  [8,] 1.0101413 -0.214854102  0.39133646 -0.539808894
#>  [9,] 1.5047888 -0.079293359  0.28603539 -0.232633615
#> [10,] 1.1645644 -0.249728464  0.24075370 -0.823428422
#> [11,] 1.4770839 -0.028411637  0.56262635 -0.465932798
#> [12,] 1.0091654 -0.444890770  0.16739869 -0.547694775
#> [13,] 1.3487066 -0.213461586  0.08268921 -0.515545997
#> [14,] 1.5631569 -0.130939629  1.00330329 -0.003436817
#> [15,] 1.1488762 -0.786734596 -0.01818211 -1.053218685
#> [16,] 1.4950159 -0.200185414  0.44138606 -0.512685055
#> [17,] 1.6763940  0.007534638  1.08889015 -0.195494246
#> [18,] 1.5982175 -0.108849758  1.05845594 -0.073666398
#> [19,] 1.5054736 -0.784471033  0.45212915 -0.261322728
#> [20,] 1.7221279  0.018575899  1.07285580  0.101293944
#> [21,] 1.6053046 -0.002283527  1.02463054 -0.255217832
#> [22,] 1.6982504 -0.273511486  0.06679731 -0.426349834
#> [23,] 1.3038393 -0.099897883  0.37800719 -0.253085755
#> [24,] 1.5085748 -0.244062635  0.47999731 -0.513198865
#> [25,] 1.0979128 -0.657044164 -0.08223721 -0.376081341
#> [26,] 1.4595504 -0.127979371  0.66936193 -0.431137069
#> [27,] 1.2187728 -0.119049070  0.27406606 -0.698539354
#> [28,] 1.6722070 -0.031078745  0.97799521 -0.248398607
#> [29,] 1.5326140 -0.370818581  0.68236685 -0.673141616
#> [30,] 0.9243895 -0.293021979 -0.21675746 -0.637866720
#> [31,] 2.1905048  0.020883179  1.14650318  0.377111626
#> [32,] 1.5711062 -0.189873477  0.95137252 -0.130121419
#> [33,] 2.0768220 -0.531852108  0.64478823  0.064156451
#> [34,] 1.6949777 -0.091147732  1.04765737  0.021987463
#> [35,] 1.2585799 -0.616949219  0.03856018 -0.751929749
#> [36,] 1.7680449  0.233313010  0.77663740 -0.086279485
#> [37,] 1.7940208 -0.089897738  0.63160165 -0.147512382
#> [38,] 1.6753522  0.222918175  0.78749891 -0.155567585
#> [39,] 1.5888493 -0.211036735  0.73546641 -0.435944104
#> [40,] 1.7538428  0.072114089  1.10091043 -0.094092817
#> 
#> $pre[[3]]
#>             [,1]        [,2]
#>  [1,] 0.04203570 -0.05356221
#>  [2,] 0.06389369 -0.05153114
#>  [3,] 0.06988861 -0.06046596
#>  [4,] 0.05328315 -0.05060547
#>  [5,] 0.05765871 -0.04727447
#>  [6,] 0.02955830 -0.04983427
#>  [7,] 0.07851324 -0.04606013
#>  [8,] 0.06197744 -0.04269653
#>  [9,] 0.05070281 -0.04190830
#> [10,] 0.07261241 -0.04796319
#> [11,] 0.05562718 -0.04900908
#> [12,] 0.06262840 -0.03861283
#> [13,] 0.06420901 -0.04131626
#> [14,] 0.03410838 -0.04727619
#> [15,] 0.09119404 -0.05319191
#> [16,] 0.05934012 -0.04986931
#> [17,] 0.04262765 -0.05485212
#> [18,] 0.03652040 -0.04953417
#> [19,] 0.05782195 -0.05200121
#> [20,] 0.02963612 -0.04753171
#> [21,] 0.04524201 -0.05401908
#> [22,] 0.06447623 -0.04685078
#> [23,] 0.05044268 -0.04043101
#> [24,] 0.06378302 -0.05365584
#> [25,] 0.06272999 -0.03496766
#> [26,] 0.05394556 -0.05051450
#> [27,] 0.06857127 -0.04745936
#> [28,] 0.04515497 -0.05452919
#> [29,] 0.06942463 -0.06222519
#> [30,] 0.07267541 -0.03254587
#> [31,] 0.01844526 -0.04783306
#> [32,] 0.04054299 -0.04942483
#> [33,] 0.03711514 -0.05040924
#> [34,] 0.03156675 -0.04956864
#> [35,] 0.07779489 -0.04672491
#> [36,] 0.03800098 -0.04560300
#> [37,] 0.04669064 -0.05179358
#> [38,] 0.04082244 -0.04600096
#> [39,] 0.05736726 -0.05680595
#> [40,] 0.03840485 -0.05311548
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4761187 -0.4761187
#>  [2,]  0.4711758 -0.4711758
#>  [3,]  0.4674574 -0.4674574
#>  [4,] -0.5259488  0.5259488
#>  [5,] -0.5262093  0.5262093
#>  [6,]  0.4801623 -0.4801623
#>  [7,]  0.4688969 -0.4688969
#>  [8,]  0.4738554 -0.4738554
#>  [9,] -0.5231362  0.5231362
#> [10,]  0.4698926 -0.4698926
#> [11,]  0.4738648 -0.4738648
#> [12,] -0.5252887  0.5252887
#> [13,]  0.4736431 -0.4736431
#> [14,]  0.4796651 -0.4796651
#> [15,] -0.5360339  0.5360339
#> [16,] -0.5272753  0.5272753
#> [17,]  0.4756493 -0.4756493
#> [18,]  0.4784996 -0.4784996
#> [19,]  0.4725718 -0.4725718
#> [20,] -0.5192824  0.5192824
#> [21,]  0.4752051 -0.4752051
#> [22,] -0.5278030  0.5278030
#> [23,]  0.4772972 -0.4772972
#> [24,]  0.4706740 -0.4706740
#> [25,]  0.4755950 -0.4755950
#> [26,]  0.4739087 -0.4739087
#> [27,]  0.4710248 -0.4710248
#> [28,]  0.4750996 -0.4750996
#> [29,]  0.4671350 -0.4671350
#> [30,]  0.4737189 -0.4737189
#> [31,] -0.5165635  0.5165635
#> [32,]  0.4775232 -0.4775232
#> [33,] -0.5218671  0.5218671
#> [34,] -0.5202727  0.5202727
#> [35,]  0.4689102 -0.4689102
#> [36,] -0.5208888  0.5208888
#> [37,] -0.5246012  0.5246012
#> [38,] -0.5216922  0.5216922
#> [39,]  0.4714877 -0.4714877
#> [40,]  0.4771359 -0.4771359
#> 
#> $L
#> [1] 0.6850535 0.7070182 0.6875233 0.6914365 0.6855580 0.6771175
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5507246