Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.2422773 -0.04231257 -0.08149239  0.01982022 -0.159815661 -0.09610674
#>  [2,] -0.3307464 -0.14269168 -0.02246518  0.03726554 -0.101866614  0.04065553
#>  [3,] -0.1961130 -0.03257635 -0.02813719 -0.06430642 -0.137604167 -0.14307346
#>  [4,] -0.2369520 -0.15714008 -0.06072456 -0.13864069 -0.049527549 -0.08801443
#>  [5,] -0.3774187 -0.13883353 -0.13449853 -0.13778983 -0.005528833 -0.11485603
#>  [6,] -0.2508747 -0.18921723 -0.21030931 -0.20412100 -0.072608845 -0.08576300
#>  [7,] -0.3064138 -0.19123269 -0.07706305 -0.12212970 -0.123001611 -0.08149749
#>  [8,] -0.3404515 -0.10936650 -0.14281869 -0.22002362 -0.105834297 -0.07108162
#>  [9,] -0.2795601 -0.03696135 -0.02778938 -0.03746316 -0.053135197 -0.01415787
#> [10,] -0.3690143 -0.05553863 -0.15392890 -0.08333447  0.015305173 -0.02313826
#>                V15         V16          V17         V18         V19         V2
#>  [1,] -0.062915507 -0.03875156  0.007126244 -0.10988315 -0.09273553 -0.2582960
#>  [2,] -0.104111970 -0.06640817 -0.043038597 -0.07053058 -0.01450075 -0.3159748
#>  [3,] -0.119348957  0.01550493 -0.028424414 -0.05609346 -0.08445578 -0.2010410
#>  [4,]  0.115920731  0.08205476 -0.038874345  0.06866472  0.10781139 -0.1808900
#>  [5,]  0.009506884  0.05038200  0.043139301  0.01674109  0.07229004 -0.3258240
#>  [6,] -0.120600894 -0.21502648 -0.069987275 -0.06676434  0.02419367 -0.2308951
#>  [7,]  0.091601805  0.05116599  0.059706089  0.13748274  0.04530717 -0.2804536
#>  [8,] -0.057756480 -0.11033482  0.036754422  0.16157999  0.14965021 -0.3651919
#>  [9,] -0.033221469  0.04823722  0.067483535  0.06272822  0.10309282 -0.3577143
#> [10,] -0.046710825  0.10874964  0.037289063  0.10668418  0.12868560 -0.2947756
#>               V20         V21        V22         V23        V24        V25
#>  [1,] -0.03477774 -0.07320402 0.03238732 -0.01999643 0.10400426 0.04518278
#>  [2,]  0.16366917  0.10794350 0.15692152  0.18884587 0.14246255 0.23144063
#>  [3,]  0.02329325  0.08263102 0.11486285  0.10797545 0.13331443 0.14188479
#>  [4,]  0.14105216  0.21706762 0.23252868  0.10811316 0.19097450 0.15790097
#>  [5,]  0.11850326 -0.01367686 0.12043615  0.06083279 0.21846776 0.09961641
#>  [6,] -0.15895846  0.04181536 0.02705503 -0.02992627 0.16488475 0.18006672
#>  [7,]  0.04759356  0.20914383 0.14042140  0.24765916 0.18528606 0.12292614
#>  [8,]  0.02132662  0.15010966 0.14783713  0.19967841 0.08918523 0.23658252
#>  [9,]  0.12152597  0.25877801 0.19615837  0.27086369 0.23521758 0.21740950
#> [10,]  0.18603149  0.27626167 0.15509362  0.27211089 0.28766314 0.18601828
#>              V26         V27         V28          V29         V3         V30
#>  [1,] 0.05019610 0.005192077  0.08931793  0.117388011 -0.2378530  0.05268254
#>  [2,] 0.19834478 0.151828788  0.17481892  0.146897590 -0.2513050  0.09371787
#>  [3,] 0.15367760 0.169694158  0.09263495  0.002342622 -0.1926836 -0.04220663
#>  [4,] 0.23290902 0.094016734  0.21493422  0.158328184 -0.1860245  0.08571115
#>  [5,] 0.07650790 0.143921370  0.15652854  0.120225682 -0.2641248  0.11521708
#>  [6,] 0.09384851 0.152469940 -0.04617014 -0.058392608 -0.3332555  0.00653453
#>  [7,] 0.16541868 0.290463582  0.24949811  0.130945564 -0.2851088  0.09385065
#>  [8,] 0.08646668 0.141534520  0.06483396  0.156107710 -0.3738783  0.09782602
#>  [9,] 0.28679768 0.305053026  0.24794985  0.179727001 -0.3680480  0.07477817
#> [10,] 0.24591676 0.132724147  0.15542559  0.096979175 -0.2767808  0.13290632
#>                 V31          V32           V33         V34         V35
#>  [1,] -7.891609e-05  0.053343830  0.0085185087 -0.06358189 -0.16405962
#>  [2,]  1.047575e-01  0.001445793  0.0549895530  0.07317244 -0.02574384
#>  [3,] -4.433923e-02 -0.010342902 -0.0494942568 -0.11946902 -0.14548688
#>  [4,] -9.595441e-03 -0.028698072  0.0324834497  0.02646728  0.01951840
#>  [5,]  6.073167e-02  0.045131358 -0.0048782801 -0.07739624 -0.02920982
#>  [6,]  2.406173e-02  0.060653998  0.0001660038 -0.06896410 -0.04751949
#>  [7,] -2.205354e-02 -0.101656680 -0.1331571833 -0.02398892 -0.07679790
#>  [8,] -3.123962e-02 -0.130359047 -0.0498768971 -0.22325153 -0.07667460
#>  [9,]  1.979473e-01  0.159462978  0.0379831350  0.05572689 -0.02714141
#> [10,]  8.914633e-02 -0.004679555 -0.0584033808  0.04847546  0.01825679
#>               V36         V37         V38         V39         V4         V40
#>  [1,] -0.06904134 -0.08618344  0.02324198 -0.05860417 -0.2769921  0.02472544
#>  [2,] -0.07774719 -0.11190878  0.02285480  0.06710401 -0.2698687 -0.06152781
#>  [3,] -0.04894544  0.01977056 -0.01301303 -0.16519942 -0.3050690 -0.13201220
#>  [4,]  0.05284660 -0.02294837 -0.05366834  0.04424313 -0.1528449 -0.06773206
#>  [5,] -0.19395510 -0.13352674 -0.22862908 -0.08329348 -0.2592575 -0.12287956
#>  [6,] -0.03337564 -0.07699464 -0.21664125 -0.11202700 -0.2050202 -0.18428115
#>  [7,] -0.08585723 -0.03994285 -0.07413616 -0.12967881 -0.2875663 -0.12286818
#>  [8,] -0.07188139 -0.14418317 -0.10015324 -0.10456114 -0.2500576 -0.18233862
#>  [9,]  0.07415493  0.03082799  0.12663885 -0.02434814 -0.1893820  0.02687186
#> [10,] -0.03133256  0.05675650  0.03068422  0.06217557 -0.2338505  0.05566314
#>               V41         V42         V43         V44         V45         V46
#>  [1,] -0.05830775 -0.12820955 -0.01280733 -0.05742848 -0.10070406 -0.19755063
#>  [2,] -0.11515339  0.05131781 -0.12764001 -0.09331248 -0.13850871 -0.11858964
#>  [3,] -0.04907403 -0.11607453 -0.19706671 -0.13838775 -0.05185276 -0.15285554
#>  [4,] -0.04198796 -0.02730760 -0.09015898 -0.06862791 -0.01949277 -0.16478614
#>  [5,] -0.06587012 -0.01846820 -0.04731017 -0.15409939 -0.12485232 -0.19776668
#>  [6,] -0.03784683 -0.10974815 -0.09570578 -0.22338130 -0.20550987 -0.13235486
#>  [7,] -0.06952475 -0.14053049 -0.01396583 -0.10369535 -0.09922855 -0.15662940
#>  [8,] -0.14065836 -0.11695607 -0.09995273 -0.07449270 -0.23567217 -0.16926927
#>  [9,] -0.04795502 -0.02193746 -0.09900325 -0.01745840 -0.06530782 -0.15141317
#> [10,]  0.03335564 -0.01004413  0.05172649 -0.03897080 -0.07723200 -0.07914336
#>              V47         V48        V49         V5        V50        V51
#>  [1,] -0.2280031 -0.16144065 -0.3178115 -0.2599418 -0.2363947 -0.2470534
#>  [2,] -0.1353086 -0.27584233 -0.1930626 -0.2739403 -0.2446051 -0.2121728
#>  [3,] -0.2266208 -0.15636738 -0.1872494 -0.1723749 -0.3089626 -0.3345962
#>  [4,] -0.1656095 -0.09710424 -0.1516368 -0.1754390 -0.3623158 -0.2991416
#>  [5,] -0.1625369 -0.16969076 -0.2722612 -0.1978820 -0.2983434 -0.3846810
#>  [6,] -0.2484498 -0.16080916 -0.2808841 -0.2500410 -0.4206026 -0.4125334
#>  [7,] -0.2034209 -0.16671524 -0.2655492 -0.2772377 -0.2842639 -0.3264351
#>  [8,] -0.1555943 -0.17065339 -0.3730702 -0.3409364 -0.3710342 -0.4120171
#>  [9,] -0.1207272 -0.23539313 -0.2219336 -0.2172583 -0.2397911 -0.3208229
#> [10,] -0.1088342 -0.29278674 -0.2907542 -0.2249124 -0.4185323 -0.3781478
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3196409 -0.3330315 -0.3223950 -0.3254375 -0.3137924 -0.3798447
#>  [2,] -0.3764285 -0.3588323 -0.2549903 -0.3731350 -0.4061202 -0.3704538
#>  [3,] -0.2952685 -0.3132843 -0.3361937 -0.2958421 -0.2784809 -0.3240922
#>  [4,] -0.3016215 -0.3380455 -0.2231679 -0.3113512 -0.3239574 -0.2113302
#>  [5,] -0.3366597 -0.2903281 -0.3621906 -0.3838542 -0.4184227 -0.4003732
#>  [6,] -0.3468320 -0.3700926 -0.3233532 -0.2601612 -0.3825454 -0.3177437
#>  [7,] -0.4111522 -0.3103897 -0.2505365 -0.2819748 -0.3919487 -0.2857398
#>  [8,] -0.2714799 -0.3394189 -0.3026098 -0.3167328 -0.2856229 -0.3451170
#>  [9,] -0.2942547 -0.2902107 -0.4161310 -0.4000651 -0.3119438 -0.3863911
#> [10,] -0.3268516 -0.4414686 -0.4259863 -0.3638261 -0.2926661 -0.3823541
#>              V58        V59          V6        V60         V7         V8
#>  [1,] -0.3878153 -0.3603634 -0.23738199 -0.3124683 -0.2881748 -0.1641497
#>  [2,] -0.3296835 -0.2969762 -0.05418793 -0.3370243 -0.1830534 -0.1913962
#>  [3,] -0.3457215 -0.2708867 -0.23672818 -0.3825927 -0.2251884 -0.2325734
#>  [4,] -0.3134397 -0.2487267 -0.08892493 -0.3361497 -0.1431773 -0.1558920
#>  [5,] -0.3887304 -0.3756299 -0.11348671 -0.3863691 -0.1263961 -0.1999339
#>  [6,] -0.3319883 -0.3334332 -0.30609214 -0.3502639 -0.1593755 -0.1111676
#>  [7,] -0.3205939 -0.3494558 -0.24484817 -0.3422810 -0.1255943 -0.1277144
#>  [8,] -0.2659132 -0.2995415 -0.26884526 -0.2560813 -0.2099176 -0.1581070
#>  [9,] -0.3112573 -0.4282390 -0.23276418 -0.3775756 -0.1925663 -0.1605143
#> [10,] -0.4445010 -0.3493809 -0.24208737 -0.3481371 -0.1838433 -0.1456090
#>                V9
#>  [1,] -0.17639288
#>  [2,] -0.11241638
#>  [3,] -0.21501439
#>  [4,] -0.12990333
#>  [5,] -0.22395768
#>  [6,] -0.10298723
#>  [7,] -0.06293595
#>  [8,] -0.12669939
#>  [9,] -0.01437824
#> [10,] -0.15148243
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,]  0.04597514  0.02075836 -0.03743727 -0.00519854  0.07011620 -0.131141225
#> [2,] -0.06150533 -0.02123950 -0.05592035 -0.15603328 -0.05575563 -0.003718464
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.10502217 -0.04103869  0.15202859  0.08696929
#> [2,] -0.08912419 -0.11542449 -0.02020009 -0.06311199
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -6.959038e-05 -7.460487e-04 -1.014787e-03 -1.092427e-03 -9.342271e-04
#>  [2,]  1.019279e-04  7.066514e-04  8.902347e-04  9.329972e-04  1.054662e-03
#>  [3,]  7.587332e-05  5.614221e-04  6.769646e-04  6.948222e-04  8.332266e-04
#>  [4,] -1.657326e-05 -1.876248e-04 -2.821704e-04 -2.969435e-04 -1.945147e-04
#>  [5,] -2.545351e-05 -2.958623e-04 -4.393986e-04 -4.775530e-04 -3.488151e-04
#>  [6,]  1.750864e-04  1.552987e-03  1.959737e-03  2.047910e-03  2.159941e-03
#>  [7,] -8.638916e-05 -7.627876e-04 -1.033902e-03 -1.076023e-03 -9.694278e-04
#>  [8,]  2.631635e-05  1.229301e-04  1.058818e-04  9.090316e-05  2.291461e-04
#>  [9,] -1.082587e-05 -9.348771e-05 -1.456265e-04 -1.500213e-04 -9.266606e-05
#> [10,]  5.588904e-06  8.833632e-06 -1.298534e-05 -1.283704e-05  5.452047e-05
#>                 V14           V15           V16           V17           V18
#>  [1,] -6.091644e-04 -4.769743e-04 -5.431929e-04 -4.979646e-04 -3.838407e-04
#>  [2,]  1.043299e-03  1.109216e-03  1.308572e-03  1.500191e-03  1.541200e-03
#>  [3,]  8.773131e-04  9.907417e-04  1.177135e-03  1.347233e-03  1.437612e-03
#>  [4,] -6.011752e-05  2.996223e-05  3.733747e-05  1.019092e-04  1.929506e-04
#>  [5,] -1.412389e-04 -1.113350e-05 -4.393280e-06  6.710024e-05  1.805294e-04
#>  [6,]  1.967380e-03  2.093279e-03  2.469425e-03  2.717910e-03  2.795261e-03
#>  [7,] -7.185376e-04 -5.940547e-04 -6.971593e-04 -6.961059e-04 -5.421651e-04
#>  [8,]  3.635093e-04  4.839252e-04  5.792762e-04  7.012114e-04  8.039674e-04
#>  [9,] -3.791303e-05  5.228675e-06  5.311421e-06  3.723590e-05  8.874881e-05
#> [10,]  1.102617e-04  1.613668e-04  1.857760e-04  2.436080e-04  2.918217e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -8.160369e-04 -1.214097e-04 -1.020025e-03 -9.645180e-04 -0.0006271186
#>  [2,]  1.850518e-03  1.577840e-04  2.002247e-03  1.995110e-03  0.0019162407
#>  [3,]  1.671237e-03  1.189940e-04  1.778711e-03  1.799873e-03  0.0018199675
#>  [4,]  5.852724e-05 -2.424033e-05 -2.875652e-05  7.387767e-06  0.0001896769
#>  [5,] -2.955521e-05 -4.166357e-05 -1.459462e-04 -9.871452e-05  0.0001394546
#>  [6,]  3.549739e-03  2.901748e-04  3.885055e-03  3.868319e-03  0.0036365868
#>  [7,] -9.540643e-04 -1.340220e-04 -1.184094e-03 -1.118037e-03 -0.0007693114
#>  [8,]  8.097318e-04  3.885902e-05  8.080072e-04  8.399061e-04  0.0009745777
#>  [9,]  2.754487e-05 -1.251724e-05 -2.394842e-05 -4.554817e-06  0.0001049869
#> [10,]  2.715447e-04  1.109496e-05  2.463447e-04  2.630685e-04  0.0003518234
#>                 V23           V24          V25           V26           V27
#>  [1,] -0.0002622782 -3.973809e-05 0.0003045054  1.874555e-04 -0.0000591573
#>  [2,]  0.0018007978  1.671722e-03 0.0014340656  1.405149e-03  0.0013568690
#>  [3,]  0.0017936860  1.722324e-03 0.0015278488  1.463702e-03  0.0013783510
#>  [4,]  0.0003680071  4.600045e-04 0.0005577846  4.699131e-04  0.0003505033
#>  [5,]  0.0003668376  4.934111e-04 0.0006534172  5.413621e-04  0.0003683518
#>  [6,]  0.0032785353  2.959149e-03 0.0023133243  2.278672e-03  0.0023398708
#>  [7,] -0.0004171914 -1.948975e-04 0.0001240825 -1.399147e-05 -0.0002085783
#>  [8,]  0.0010855915  1.121056e-03 0.0011144591  1.022445e-03  0.0008810168
#>  [9,]  0.0002170255  2.812969e-04 0.0003341337  2.863460e-04  0.0002220751
#> [10,]  0.0004334554  4.696215e-04 0.0004898584  4.441746e-04  0.0003805709
#>                 V28           V29            V3           V30          V31
#>  [1,] -0.0002498343 -0.0001975029 -1.294120e-04  9.227996e-06 0.0003226224
#>  [2,]  0.0014708805  0.0013893407  1.766363e-04  1.263729e-03 0.0010608647
#>  [3,]  0.0014744139  0.0013941842  1.379115e-04  1.280834e-03 0.0010886152
#>  [4,]  0.0003081910  0.0003022978 -2.616345e-05  3.371686e-04 0.0004277172
#>  [5,]  0.0002935430  0.0002919877 -4.414194e-05  3.732275e-04 0.0005267329
#>  [6,]  0.0026925419  0.0025316287  3.322868e-04  2.168742e-03 0.0015513756
#>  [7,] -0.0003584774 -0.0003068796 -1.503656e-04 -1.652948e-04 0.0001257398
#>  [8,]  0.0008870526  0.0008493410  4.729807e-05  8.450121e-04 0.0008358619
#>  [9,]  0.0002093254  0.0002035617 -1.528873e-05  2.118363e-04 0.0002520934
#> [10,]  0.0003823085  0.0003689916  1.085196e-05  3.636643e-04 0.0003821166
#>                V32          V33          V34          V35           V36
#>  [1,] 0.0003272411 0.0005277233 0.0007359521 0.0010166123  0.0013533939
#>  [2,] 0.0009638888 0.0007528057 0.0006684784 0.0005237015  0.0001564630
#>  [3,] 0.0009520955 0.0007934790 0.0007379398 0.0006514689  0.0003922858
#>  [4,] 0.0003956744 0.0004380351 0.0004986729 0.0006274862  0.0007698971
#>  [5,] 0.0004974220 0.0005671362 0.0006603617 0.0008336534  0.0010111214
#>  [6,] 0.0012820135 0.0008466024 0.0005830820 0.0002049027 -0.0005225150
#>  [7,] 0.0001533610 0.0003717015 0.0005400430 0.0008457231  0.0012881786
#>  [8,] 0.0007509811 0.0007057034 0.0007369100 0.0007885119  0.0007433839
#>  [9,] 0.0002298564 0.0002444465 0.0002693964 0.0003335266  0.0004256179
#> [10,] 0.0003506516 0.0003339131 0.0003505315 0.0003977082  0.0004247696
#>                 V37           V38           V39            V4          V40
#>  [1,]  9.184333e-04 -1.221206e-06 -3.270807e-04 -2.117953e-04 0.0002960390
#>  [2,]  2.678310e-04  6.829506e-04  8.125730e-04  2.510718e-04 0.0003593662
#>  [3,]  4.597792e-04  6.605826e-04  7.100080e-04  1.958785e-04 0.0004252069
#>  [4,]  6.020818e-04  2.017981e-04  6.522875e-05 -5.265889e-05 0.0003055098
#>  [5,]  7.701451e-04  2.007723e-04  5.813810e-06 -8.031033e-05 0.0003573291
#>  [6,] -2.035322e-05  1.078219e-03  1.403437e-03  5.007574e-04 0.0004130395
#>  [7,]  9.000529e-04 -6.739790e-05 -3.874443e-04 -2.373687e-04 0.0003168179
#>  [8,]  6.406668e-04  4.408670e-04  3.639918e-04  5.647919e-05 0.0004000303
#>  [9,]  3.383590e-04  1.248851e-04  6.290400e-05 -3.081556e-05 0.0001935185
#> [10,]  3.534977e-04  2.055459e-04  1.591563e-04  6.408250e-06 0.0002163160
#>                V41           V42           V43           V44           V45
#>  [1,] 0.0001979963 -2.168909e-04 -5.154252e-04 -6.803880e-04 -8.583231e-04
#>  [2,] 0.0003708651  5.321408e-04  7.162742e-04  7.309572e-04  7.611044e-04
#>  [3,] 0.0004000459  4.251715e-04  5.617517e-04  5.670155e-04  5.565625e-04
#>  [4,] 0.0002507100  4.693467e-05 -7.295441e-05 -1.467573e-04 -2.507689e-04
#>  [5,] 0.0002861421 -1.338937e-06 -1.638765e-04 -2.633363e-04 -4.002255e-04
#>  [6,] 0.0004459446  7.889827e-04  1.297534e-03  1.466240e-03  1.594131e-03
#>  [7,] 0.0002077569 -2.467112e-04 -5.572920e-04 -7.012440e-04 -9.248188e-04
#>  [8,] 0.0003482384  2.132631e-04  2.014426e-04  1.495479e-04  7.224157e-05
#>  [9,] 0.0001637011  5.000092e-05 -1.955775e-05 -5.569102e-05 -1.108224e-04
#> [10,] 0.0001912498  1.114645e-04  7.434067e-05  4.076900e-05 -7.908230e-06
#>                 V46           V47           V48           V49            V5
#>  [1,] -6.505156e-04 -4.004330e-04 -3.328004e-04 -2.111976e-04 -2.432785e-04
#>  [2,]  6.259641e-04  4.491926e-04  3.551061e-04  2.160332e-04  3.048909e-04
#>  [3,]  4.576858e-04  3.506308e-04  2.721659e-04  1.672564e-04  2.501738e-04
#>  [4,] -1.877879e-04 -8.222455e-05 -8.056665e-05 -4.994992e-05 -4.676702e-05
#>  [5,] -2.957621e-04 -1.462322e-04 -1.317217e-04 -8.430197e-05 -8.225937e-05
#>  [6,]  1.268286e-03  9.099212e-04  7.213661e-04  4.501963e-04  6.203112e-04
#>  [7,] -7.247152e-04 -4.293895e-04 -3.676974e-04 -2.264227e-04 -2.643656e-04
#>  [8,]  8.048235e-05  9.902961e-05  6.685734e-05  3.839455e-05  8.117797e-05
#>  [9,] -8.249828e-05 -2.813339e-05 -3.167861e-05 -1.881349e-05 -2.235655e-05
#> [10,]  3.195632e-06  3.119585e-05  1.542578e-05  9.266104e-06  1.910553e-05
#>                 V50           V51           V52           V53           V54
#>  [1,] -4.179916e-05 -5.678255e-05 -4.317571e-05 -2.423491e-05 -2.108728e-05
#>  [2,]  6.322844e-05  6.602710e-05  5.130708e-05  3.502848e-05  3.548392e-05
#>  [3,]  5.046563e-05  4.796362e-05  3.712383e-05  2.708172e-05  2.776029e-05
#>  [4,] -6.992345e-06 -1.389481e-05 -1.112751e-05 -4.560909e-06 -3.153634e-06
#>  [5,] -1.343752e-05 -2.273665e-05 -1.700073e-05 -7.840115e-06 -5.945545e-06
#>  [6,]  1.156122e-04  1.247930e-04  9.566303e-05  6.257729e-05  6.087887e-05
#>  [7,] -5.046042e-05 -6.349745e-05 -4.851691e-05 -2.788069e-05 -2.512375e-05
#>  [8,]  1.841153e-05  1.230853e-05  9.933672e-06  9.670670e-06  1.122168e-05
#>  [9,] -2.201537e-06 -6.214769e-06 -5.872773e-06 -2.685393e-06 -1.560636e-06
#> [10,]  5.907079e-06  2.938550e-06  1.715447e-06  2.656653e-06  3.357831e-06
#>                 V55           V56          V57           V58           V59
#>  [1,] -1.679122e-05 -1.545471e-05 3.614506e-06 -1.897728e-05 -2.162351e-05
#>  [2,]  3.430248e-05  3.059588e-05 1.608755e-05  3.638502e-05  3.184742e-05
#>  [3,]  2.704764e-05  2.582590e-05 1.579054e-05  2.985293e-05  2.421885e-05
#>  [4,] -1.801573e-06 -9.858487e-07 5.590985e-06 -1.607796e-06 -4.093496e-06
#>  [5,] -4.112955e-06 -2.515405e-06 7.203742e-06 -3.937315e-06 -7.350397e-06
#>  [6,]  5.726863e-05  5.550035e-05 2.256074e-05  6.490942e-05  5.667708e-05
#>  [7,] -2.338761e-05 -1.910557e-05 8.756658e-07 -2.427534e-05 -2.585968e-05
#>  [8,]  1.183961e-05  1.178112e-05 1.200729e-05  1.330340e-05  8.572954e-06
#>  [9,] -9.020237e-07 -6.124598e-07 2.470175e-06 -8.046477e-07 -2.414290e-06
#> [10,]  3.996759e-06  3.661393e-06 4.887359e-06  4.308508e-06  2.340614e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -1.477603e-04 -8.698473e-06 -1.088706e-04 -2.140893e-04 -5.976389e-04
#>  [2,]  2.703380e-04  1.880526e-05  2.772873e-04  3.024979e-04  5.682051e-04
#>  [3,]  2.382837e-04  1.552896e-05  2.518204e-04  2.446116e-04  4.434476e-04
#>  [4,]  3.492522e-06 -2.246157e-07  2.970973e-05 -2.571567e-05 -1.621652e-04
#>  [5,] -1.473814e-05 -1.168024e-06  1.403848e-05 -5.382369e-05 -2.486726e-04
#>  [6,]  5.197523e-04  3.241294e-05  5.078848e-04  5.680048e-04  1.225977e-03
#>  [7,] -1.513144e-04 -1.106100e-05 -1.137568e-04 -2.215799e-04 -6.344337e-04
#>  [8,]  1.081038e-04  7.294561e-06  1.289027e-04  9.085854e-05  9.172738e-05
#>  [9,]  7.372897e-06 -2.648944e-07  2.356922e-05 -1.178263e-05 -8.371695e-05
#> [10,]  4.225775e-05  2.394329e-06  5.761967e-05  3.168711e-05 -1.116646e-06
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]       [,6]
#> [1,] -0.02494116 -0.06441448 -0.03668401 -0.0644948 -0.05103721 -0.0164258
#> [2,]  0.02494116  0.06441448  0.03668401  0.0644948  0.05103721  0.0164258
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.05884989 -0.04562592 -0.07553702 -0.07479948
#> [2,]  0.05884989  0.04562592  0.07553702  0.07479948
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.1957751 0.6328728 0.3398474 0.2962765 0.4624181 0.2905652 0.3063643
#>  [8] 0.3228129 0.5297347 0.6614648
#> 
#> $B[[2]]
#> [1]  0.1907646 -0.1618758
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -0.0001710240  0.0023874594  0.0022990110  0.0005363111  0.0005648821
#>  [6]  0.0039733110 -0.0004538557  0.0014501844  0.0003116936  0.0006114520
#> 
#> $vB[[2]]
#> [1] -0.08802835  0.08802835
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#>  [2,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#>  [3,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [4,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#>  [5,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#>  [6,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [7,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#>  [8,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#>  [9,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [10,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [11,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [12,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [13,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [14,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [15,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [16,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [17,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [18,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [19,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [20,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [21,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [22,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [23,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [24,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [25,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [26,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [27,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [28,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [29,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [30,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [31,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [32,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [33,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [34,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [35,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [36,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [37,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [38,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [39,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [40,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#>  [2,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#>  [3,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [4,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#>  [5,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#>  [6,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [7,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#>  [8,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#>  [9,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [10,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [11,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [12,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [13,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [14,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [15,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [16,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [17,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [18,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [19,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [20,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [21,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [22,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [23,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [24,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [25,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [26,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [27,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [28,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [29,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [30,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [31,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [32,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [33,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [34,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [35,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [36,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [37,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [38,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [39,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [40,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#>  [2,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#>  [3,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [4,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#>  [5,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#>  [6,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [7,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#>  [8,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#>  [9,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [10,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [11,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [12,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [13,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [14,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [15,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [16,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [17,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [18,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [19,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [20,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [21,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [22,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [23,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [24,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [25,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [26,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [27,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [28,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [29,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [30,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [31,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [32,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [33,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [34,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [35,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [36,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [37,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [38,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [39,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [40,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#>  [2,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#>  [3,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [4,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#>  [5,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#>  [6,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [7,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#>  [8,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#>  [9,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [10,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [11,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [12,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [13,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [14,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [15,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [16,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [17,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [18,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [19,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [20,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [21,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [22,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [23,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [24,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [25,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [26,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [27,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [28,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [29,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [30,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [31,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [32,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [33,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [34,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [35,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [36,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [37,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [38,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [39,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [40,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#>  [2,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#>  [3,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [4,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#>  [5,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#>  [6,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [7,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#>  [8,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#>  [9,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [10,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [11,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [12,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [13,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [14,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [15,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [16,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [17,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [18,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [19,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [20,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [21,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [22,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [23,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [24,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [25,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [26,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [27,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [28,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [29,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [30,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [31,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [32,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [33,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [34,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [35,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [36,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [37,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [38,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [39,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [40,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#>  [2,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#>  [3,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [4,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#>  [5,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#>  [6,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [7,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#>  [8,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#>  [9,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [10,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [11,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [12,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [13,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [14,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [15,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [16,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [17,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [18,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [19,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [20,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [21,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [22,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [23,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [24,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [25,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [26,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [27,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [28,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [29,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [30,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [31,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [32,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [33,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [34,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [35,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [36,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [37,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [38,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [39,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [40,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3737113 0.7963653 0.5123583 0.8043256 0.6126538 0.3876621 0.6884119
#>  [2,] 0.5003144 0.8581583 0.6446954 0.8281370 0.7402942 0.5066466 0.8085965
#>  [3,] 0.4893437 0.8432428 0.5965973 0.8146948 0.6728342 0.4979546 0.7454126
#>  [4,] 0.4259025 0.7801773 0.5719176 0.7899114 0.6509961 0.4182522 0.7427010
#>  [5,] 0.4801008 0.8471664 0.6307899 0.8240093 0.7358689 0.4773488 0.8097374
#>  [6,] 0.4148141 0.7806649 0.5267177 0.7999421 0.6934388 0.4040769 0.7433854
#>  [7,] 0.5238148 0.8489149 0.6565807 0.8214314 0.7397213 0.5382180 0.8038831
#>  [8,] 0.5127370 0.8745868 0.6334437 0.8496618 0.7231007 0.5249535 0.7967320
#>  [9,] 0.3780207 0.7370701 0.4311842 0.7116425 0.5438635 0.3219998 0.6035425
#> [10,] 0.4293730 0.8159349 0.5647623 0.7894525 0.6272497 0.3827474 0.7353932
#> [11,] 0.3480952 0.7713867 0.4927881 0.7966381 0.6172771 0.3533176 0.7146983
#> [12,] 0.3342261 0.6504344 0.4150796 0.6830887 0.5511011 0.3066423 0.5894089
#> [13,] 0.4522511 0.8352709 0.6092570 0.8340069 0.7258325 0.4467228 0.8134726
#> [14,] 0.4986832 0.8237173 0.5896310 0.7727712 0.7053915 0.4911051 0.7423784
#> [15,] 0.4445518 0.8243543 0.5653542 0.7902505 0.6569263 0.4776109 0.7195494
#> [16,] 0.4307938 0.8140195 0.5757411 0.8316523 0.7309520 0.4174359 0.8070715
#> [17,] 0.4734888 0.7958782 0.5568964 0.7634743 0.6630829 0.4359158 0.7242457
#> [18,] 0.3659536 0.7200828 0.4562071 0.7374009 0.6087844 0.3438419 0.6536277
#> [19,] 0.3797582 0.7586637 0.4779193 0.7284484 0.5418070 0.3254412 0.6292935
#> [20,] 0.3014154 0.6461283 0.3130900 0.6105103 0.4652015 0.2441426 0.4768168
#> [21,] 0.4351118 0.7954839 0.5416417 0.7833697 0.5814329 0.3783688 0.7085242
#> [22,] 0.4198319 0.7775233 0.4624553 0.7305892 0.5680447 0.3574949 0.6227166
#> [23,] 0.5528628 0.8646831 0.6829054 0.8275709 0.7719873 0.5850443 0.8212672
#> [24,] 0.3766993 0.7750603 0.5035640 0.7944029 0.6760600 0.3559682 0.7387136
#> [25,] 0.3567822 0.7700795 0.4550033 0.7636430 0.6157060 0.3234082 0.6662669
#> [26,] 0.5417150 0.8596921 0.6485958 0.8104529 0.7472326 0.5419929 0.7831541
#> [27,] 0.3081834 0.7528721 0.4043697 0.7247822 0.5681098 0.2858121 0.6118833
#> [28,] 0.3348325 0.7319680 0.4412752 0.7529194 0.6034456 0.2882268 0.6884271
#> [29,] 0.4444588 0.8457091 0.5970106 0.8421485 0.6522927 0.4280531 0.7795737
#> [30,] 0.4061585 0.6899151 0.3847664 0.6464262 0.4820777 0.3295148 0.4971379
#> [31,] 0.5245507 0.8068699 0.6283031 0.7589706 0.6951618 0.5551620 0.7394640
#> [32,] 0.4920898 0.8606912 0.6304030 0.8323430 0.7000370 0.4844176 0.7925311
#> [33,] 0.3656750 0.7372045 0.4966009 0.7648512 0.5775585 0.3503803 0.7073527
#> [34,] 0.5167788 0.8222480 0.6321099 0.7884603 0.6910353 0.5267024 0.7545059
#> [35,] 0.4039747 0.7842509 0.5300253 0.8023814 0.6690387 0.3717116 0.7475631
#> [36,] 0.4139923 0.8078016 0.5588290 0.8310627 0.7273961 0.4099106 0.7921918
#> [37,] 0.4276850 0.7760757 0.4661016 0.7225218 0.5766442 0.3802615 0.6279509
#> [38,] 0.4447893 0.8139509 0.5125638 0.7946043 0.6149007 0.4116811 0.7066066
#> [39,] 0.4583268 0.8495471 0.6154908 0.8439302 0.7435725 0.4470579 0.8194059
#> [40,] 0.3787040 0.7846180 0.5083592 0.7957704 0.6284345 0.3758696 0.7007006
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5252696 0.9154254 0.8988246
#>  [2,] 0.7081972 0.9268415 0.9053446
#>  [3,] 0.6183940 0.9288722 0.9004045
#>  [4,] 0.6407924 0.8955619 0.8865287
#>  [5,] 0.7060856 0.9234616 0.9059112
#>  [6,] 0.6583117 0.8867727 0.8857361
#>  [7,] 0.7112643 0.9214588 0.9023454
#>  [8,] 0.6754401 0.9487044 0.9227829
#>  [9,] 0.4197305 0.8640769 0.8345909
#> [10,] 0.5810224 0.9162103 0.9000956
#> [11,] 0.5951892 0.8915865 0.8955186
#> [12,] 0.4869060 0.7875910 0.7946805
#> [13,] 0.6998498 0.9313842 0.9172932
#> [14,] 0.6337871 0.8973503 0.8658062
#> [15,] 0.6081039 0.9097046 0.8973675
#> [16,] 0.7029700 0.9163529 0.9079784
#> [17,] 0.5765946 0.8962239 0.8717443
#> [18,] 0.5536315 0.8463285 0.8572122
#> [19,] 0.4789872 0.8698434 0.8549206
#> [20,] 0.3290118 0.7733479 0.7426648
#> [21,] 0.5612649 0.9002674 0.8814117
#> [22,] 0.4352163 0.9005249 0.8489927
#> [23,] 0.7504135 0.9286708 0.9083354
#> [24,] 0.6188249 0.8949892 0.8872712
#> [25,] 0.4838452 0.8952204 0.8705840
#> [26,] 0.6811760 0.9281669 0.8908064
#> [27,] 0.4543643 0.8856327 0.8633005
#> [28,] 0.5379248 0.8651306 0.8648130
#> [29,] 0.6401347 0.9404086 0.9235826
#> [30,] 0.3462475 0.8093930 0.7794596
#> [31,] 0.6516850 0.8695587 0.8539406
#> [32,] 0.6699166 0.9383820 0.9182420
#> [33,] 0.5418846 0.8916045 0.8725280
#> [34,] 0.6583688 0.8950489 0.8806962
#> [35,] 0.6391394 0.8999625 0.9002632
#> [36,] 0.6968372 0.9106088 0.9059366
#> [37,] 0.4533080 0.8956629 0.8428431
#> [38,] 0.5401161 0.9299267 0.8991794
#> [39,] 0.7077762 0.9339736 0.9200399
#> [40,] 0.5713429 0.8992375 0.8991831
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5335553 0.4664447
#>  [2,] 0.5299444 0.4700556
#>  [3,] 0.5298253 0.4701747
#>  [4,] 0.5332101 0.4667899
#>  [5,] 0.5312159 0.4687841
#>  [6,] 0.5344224 0.4655776
#>  [7,] 0.5288871 0.4711129
#>  [8,] 0.5292343 0.4707657
#>  [9,] 0.5362643 0.4637357
#> [10,] 0.5340801 0.4659199
#> [11,] 0.5354309 0.4645691
#> [12,] 0.5372696 0.4627304
#> [13,] 0.5327253 0.4672747
#> [14,] 0.5304225 0.4695775
#> [15,] 0.5300526 0.4699474
#> [16,] 0.5343054 0.4656946
#> [17,] 0.5329015 0.4670985
#> [18,] 0.5360702 0.4639298
#> [19,] 0.5354006 0.4645994
#> [20,] 0.5383133 0.4616867
#> [21,] 0.5342759 0.4657241
#> [22,] 0.5348869 0.4651131
#> [23,] 0.5271626 0.4728374
#> [24,] 0.5362188 0.4637812
#> [25,] 0.5366654 0.4633346
#> [26,] 0.5286196 0.4713804
#> [27,] 0.5370760 0.4629240
#> [28,] 0.5386674 0.4613326
#> [29,] 0.5325627 0.4674373
#> [30,] 0.5352906 0.4647094
#> [31,] 0.5276570 0.4723430
#> [32,] 0.5305559 0.4694441
#> [33,] 0.5361595 0.4638405
#> [34,] 0.5288294 0.4711706
#> [35,] 0.5355348 0.4644652
#> [36,] 0.5344180 0.4655820
#> [37,] 0.5340774 0.4659226
#> [38,] 0.5335539 0.4664461
#> [39,] 0.5326580 0.4673420
#> [40,] 0.5344078 0.4655922
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]      [,2]        [,3]      [,4]        [,5]         [,6]
#>  [1,] -0.516327716 1.3637305  0.04944331 1.4135518  0.45848105 -0.457150134
#>  [2,]  0.001257540 1.8000769  0.59580233 1.5724809  1.04749845  0.026587871
#>  [3,] -0.042631773 1.6825567  0.39130715 1.4808091  0.72103166 -0.008181679
#>  [4,] -0.298588855 1.2667001  0.28967936 1.3243915  0.62342037 -0.329952405
#>  [5,] -0.079639019 1.7125472  0.53560692 1.5437509  1.02460628 -0.090666844
#>  [6,] -0.344099081 1.2695455  0.10697251 1.3859326  0.81624578 -0.388506314
#>  [7,]  0.095331231 1.7261154  0.64809331 1.5260755  1.04452065  0.153170707
#>  [8,]  0.050958929 1.9421374  0.54701904 1.7319510  0.95989474  0.099897078
#>  [9,] -0.497957752 1.0307957 -0.27702111 0.9033745  0.17590629 -0.744596809
#> [10,] -0.284409563 1.4890450  0.26051269 1.3216283  0.52043588 -0.477903240
#> [11,] -0.627422439 1.2161581 -0.02884972 1.3654133  0.47800667 -0.604487780
#> [12,] -0.689132324 0.6209490 -0.34300558 0.7680030  0.20512068 -0.815864056
#> [13,] -0.191579594 1.6234535  0.44419026 1.6142952  0.97357998 -0.213920963
#> [14,] -0.005267241 1.5417381  0.36244033 1.2240255  0.87310557 -0.035583527
#> [15,] -0.222708606 1.5461312  0.26292080 1.3264362  0.64962664 -0.089616383
#> [16,] -0.278613357 1.4763428  0.30531407 1.5973829  0.99945779 -0.333308225
#> [17,] -0.106144426 1.3607295  0.22857578 1.1718226  0.67706334 -0.257754261
#> [18,] -0.549615572 0.9448725 -0.17562174 1.0325030  0.44220533 -0.646219599
#> [19,] -0.490574878 1.1453674 -0.08838023 0.9867647  0.16761934 -0.728876889
#> [20,] -0.840566938 0.6020637 -0.78571260 0.4494576 -0.13941956 -1.130100337
#> [21,] -0.261024994 1.3583037  0.16695340 1.2854123  0.32865834 -0.496477618
#> [22,] -0.323463368 1.2512911 -0.15046194 0.9976140  0.27387807 -0.586253671
#> [23,]  0.212244489 1.8547436  0.76715623 1.5685087  1.21956661  0.343515641
#> [24,] -0.503581774 1.2371084  0.01425639 1.3516725  0.73572331 -0.592906728
#> [25,] -0.589358045 1.2087599 -0.18047507 1.1727573  0.47136179 -0.738152993
#> [26,]  0.167248654 1.8127349  0.61287257 1.4529556  1.08390662  0.168368212
#> [27,] -0.808625840 1.1139893 -0.38729058 0.9683082  0.27414339 -0.915811430
#> [28,] -0.686408467 1.0046302 -0.23598841 1.1142434  0.41984268 -0.904011707
#> [29,] -0.223085215 1.7013354  0.39302464 1.6743013  0.62913250 -0.289798772
#> [30,] -0.379868868 0.7997226 -0.46936551 0.6033669 -0.07171973 -0.710380401
#> [31,]  0.098281761 1.4297986  0.52494380 1.1470439  0.82436351  0.221549786
#> [32,] -0.031643519 1.8210427  0.53394602 1.6023243  0.84747395 -0.062349855
#> [33,] -0.550816663 1.0314893 -0.01359674 1.1794628  0.31275895 -0.617367900
#> [34,]  0.067140579 1.5316530  0.54127894 1.3156696  0.80496393  0.106911180
#> [35,] -0.388930861 1.2906128  0.12024568 1.4012452  0.70384034 -0.524881106
#> [36,] -0.347485661 1.4357883  0.23641104 1.5931776  0.98145154 -0.364334857
#> [37,] -0.291302753 1.2429422 -0.13580182 0.9570050  0.30901228 -0.488438616
#> [38,] -0.221746923 1.4758891  0.05026596 1.3529061  0.46795954 -0.357020063
#> [39,] -0.167080586 1.7310532  0.47045232 1.6877665  1.06462011 -0.212565282
#> [40,] -0.495052888 1.2927837  0.03344009 1.3600658  0.52550659 -0.507116992
#>              [,7]        [,8]     [,9]    [,10]
#>  [1,]  0.79270510  0.10116440 2.381755 2.184233
#>  [2,]  1.44091651  0.88664456 2.539154 2.258073
#>  [3,]  1.07429380  0.48273695 2.569492 2.201727
#>  [4,]  1.06005478  0.57880507 2.148857 2.055764
#>  [5,]  1.44830497  0.87644787 2.490336 2.264702
#>  [6,]  1.06363906  0.65577949 2.058192 2.047909
#>  [7,]  1.41074321  0.90153247 2.462335 2.223561
#>  [8,]  1.36599279  0.73289457 2.917491 2.480773
#>  [9,]  0.42024755 -0.32387962 1.849573 1.618520
#> [10,]  1.02216018  0.32697196 2.391936 2.198288
#> [11,]  0.91831330  0.38545950 2.107050 2.148394
#> [12,]  0.36152229 -0.05238781 1.310465 1.353373
#> [13,]  1.47273404  0.84658295 2.608149 2.406125
#> [14,]  1.05836712  0.54849844 2.168124 1.864376
#> [15,]  0.94222738  0.43934886 2.310033 2.168310
#> [16,]  1.43109255  0.86148105 2.393795 2.289197
#> [17,]  0.96562061  0.30880945 2.155954 1.916470
#> [18,]  0.63502356  0.21535432 1.706090 1.792326
#> [19,]  0.52918727 -0.08410057 1.899575 1.773728
#> [20,] -0.09279921 -0.71265775 1.227313 1.059865
#> [21,]  0.88822720  0.24629733 2.200199 2.005867
#> [22,]  0.50109483 -0.26059976 2.203071 1.726723
#> [23,]  1.52495687  1.10081887 2.566448 2.293477
#> [24,]  1.03929324  0.48456352 2.142749 2.063166
#> [25,]  0.69134860 -0.06464172 2.145211 1.906133
#> [26,]  1.28414250  0.75918145 2.558866 2.099005
#> [27,]  0.45523531 -0.18305219 2.046887 1.842978
#> [28,]  0.79277632  0.15199121 1.858574 1.855854
#> [29,]  1.26318380  0.57594896 2.758803 2.492050
#> [30,] -0.01144847 -0.63557480 1.446071 1.262520
#> [31,]  1.04318442  0.62645397 1.897063 1.765849
#> [32,]  1.34025002  0.70780773 2.723204 2.418697
#> [33,]  0.88256149  0.16793194 2.107236 1.923498
#> [34,]  1.12279031  0.65603346 2.143384 1.999040
#> [35,]  1.08565732  0.57163065 2.196808 2.200152
#> [36,]  1.33818806  0.83228180 2.321091 2.265000
#> [37,]  0.52343602 -0.18731383 2.149937 1.679536
#> [38,]  0.87895979  0.16081002 2.585563 2.188140
#> [39,]  1.51232786  0.88460797 2.649393 2.442890
#> [40,]  0.85063632  0.28733215 2.188781 2.188180
#> 
#> $pre[[3]]
#>                [,1]       [,2]
#>  [1,]  0.0035046392 -0.1309185
#>  [2,] -0.0324730384 -0.1523940
#>  [3,] -0.0225945104 -0.1420377
#>  [4,] -0.0110575877 -0.1440940
#>  [5,] -0.0268769890 -0.1519031
#>  [6,] -0.0079279799 -0.1458358
#>  [7,] -0.0375903102 -0.1532676
#>  [8,] -0.0324490185 -0.1495199
#>  [9,]  0.0275555926 -0.1177567
#> [10,] -0.0006099373 -0.1371422
#> [11,]  0.0044893890 -0.1374723
#> [12,]  0.0231940941 -0.1261615
#> [13,] -0.0202588699 -0.1513476
#> [14,] -0.0207819406 -0.1426225
#> [15,] -0.0177996212 -0.1381552
#> [16,] -0.0145437412 -0.1519812
#> [17,] -0.0055799225 -0.1373762
#> [18,]  0.0120346652 -0.1324971
#> [19,]  0.0182090589 -0.1236309
#> [20,]  0.0496565017 -0.1038976
#> [21,]  0.0013380327 -0.1359810
#> [22,]  0.0207606261 -0.1190140
#> [23,] -0.0483068486 -0.1570643
#> [24,]  0.0042384772 -0.1408908
#> [25,]  0.0216009579 -0.1253243
#> [26,] -0.0347178888 -0.1493217
#> [27,]  0.0316530752 -0.1169237
#> [28,]  0.0234902332 -0.1314888
#> [29,] -0.0148093622 -0.1452448
#> [30,]  0.0325055851 -0.1088918
#> [31,] -0.0350107232 -0.1457516
#> [32,] -0.0258626605 -0.1482389
#> [33,]  0.0119618528 -0.1329292
#> [34,] -0.0316713144 -0.1471170
#> [35,] -0.0013770100 -0.1437563
#> [36,] -0.0129081719 -0.1507981
#> [37,]  0.0163124779 -0.1202088
#> [38,]  0.0020596631 -0.1323581
#> [39,] -0.0218004376 -0.1526186
#> [40,]  0.0023970090 -0.1354520
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5335553  0.5335553
#>  [2,]  0.4700556 -0.4700556
#>  [3,] -0.5298253  0.5298253
#>  [4,]  0.4667899 -0.4667899
#>  [5,]  0.4687841 -0.4687841
#>  [6,]  0.4655776 -0.4655776
#>  [7,] -0.5288871  0.5288871
#>  [8,] -0.5292343  0.5292343
#>  [9,] -0.5362643  0.5362643
#> [10,]  0.4659199 -0.4659199
#> [11,]  0.4645691 -0.4645691
#> [12,]  0.4627304 -0.4627304
#> [13,]  0.4672747 -0.4672747
#> [14,]  0.4695775 -0.4695775
#> [15,] -0.5300526  0.5300526
#> [16,]  0.4656946 -0.4656946
#> [17,]  0.4670985 -0.4670985
#> [18,]  0.4639298 -0.4639298
#> [19,]  0.4645994 -0.4645994
#> [20,]  0.4616867 -0.4616867
#> [21,]  0.4657241 -0.4657241
#> [22,]  0.4651131 -0.4651131
#> [23,] -0.5271626  0.5271626
#> [24,] -0.5362188  0.5362188
#> [25,]  0.4633346 -0.4633346
#> [26,]  0.4713804 -0.4713804
#> [27,]  0.4629240 -0.4629240
#> [28,]  0.4613326 -0.4613326
#> [29,]  0.4674373 -0.4674373
#> [30,] -0.5352906  0.5352906
#> [31,] -0.5276570  0.5276570
#> [32,]  0.4694441 -0.4694441
#> [33,] -0.5361595  0.5361595
#> [34,] -0.5288294  0.5288294
#> [35,]  0.4644652 -0.4644652
#> [36,]  0.4655820 -0.4655820
#> [37,]  0.4659226 -0.4659226
#> [38,] -0.5335539  0.5335539
#> [39,]  0.4673420 -0.4673420
#> [40,]  0.4655922 -0.4655922
#> 
#> $L
#> [1] 0.7125580 0.8762492 0.7033595 0.6655377 0.6969979 0.6699756
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.6521739