Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10           V11         V12          V13
#>  [1,] -0.2482194 -0.14032174 -0.0128409830 -0.14613266  0.006177990
#>  [2,] -0.3913516 -0.05473083 -0.0019160432 -0.06062082 -0.044422651
#>  [3,] -0.2842567 -0.12759685 -0.1257850796 -0.09100013 -0.098427750
#>  [4,] -0.3632366  0.04466481 -0.1346969602 -0.01776811 -0.083930928
#>  [5,] -0.2544466 -0.12346729 -0.0493907645 -0.17511428 -0.078104440
#>  [6,] -0.2331612 -0.16693295 -0.0378555416 -0.15286483 -0.015762323
#>  [7,] -0.2377428 -0.13609059 -0.0002337217 -0.02257044  0.034609932
#>  [8,] -0.3888387 -0.08586961 -0.0989633583 -0.23672799 -0.134683577
#>  [9,] -0.2619298 -0.08227961 -0.1348721515 -0.05113701  0.007832199
#> [10,] -0.3229948 -0.16573692 -0.1695700347 -0.01573447 -0.052746456
#>                V14          V15          V16         V17          V18
#>  [1,] -0.133246545 -0.068527259 -0.077769402 -0.03143650  0.083562582
#>  [2,] -0.074003018  0.034304534  0.095093031  0.03318086  0.036166467
#>  [3,] -0.012583295 -0.045878300 -0.025243149  0.13740303  0.009469293
#>  [4,] -0.037730492 -0.140631969  0.086082669  0.07198488  0.042289168
#>  [5,] -0.121111972 -0.043996636 -0.048504990 -0.02669083 -0.090870809
#>  [6,]  0.010288613 -0.053240707 -0.018501256  0.02883112  0.083302790
#>  [7,]  0.027154498 -0.007746794 -0.003860602 -0.02252072  0.100494428
#>  [8,] -0.028006444 -0.089135973 -0.121894524 -0.08211159  0.035918581
#>  [9,] -0.125511855  0.045966766 -0.053607701  0.09760269 -0.024945134
#> [10,]  0.005248289 -0.062256623  0.074213380 -0.01345543  0.007712130
#>               V19         V2          V20         V21         V22         V23
#>  [1,]  0.05608595 -0.2039931 0.0015861200  0.11391231  0.16195121  0.12569927
#>  [2,]  0.06767652 -0.4089522 0.2110484049  0.24153433  0.17673470  0.24735913
#>  [3,]  0.09777885 -0.2623443 0.0561306414  0.07385092  0.25648458  0.13859481
#>  [4,]  0.14479099 -0.3772969 0.0878252970  0.17783184  0.14673682  0.20079877
#>  [5,] -0.06915122 -0.1506198 0.1295214156  0.14554367 -0.01623545 -0.02680453
#>  [6,]  0.05305698 -0.2734210 0.0276899105 -0.03894326  0.17228301  0.04617843
#>  [7,]  0.05719585 -0.3398365 0.0440616288  0.17971570  0.06013105  0.10375953
#>  [8,]  0.03114461 -0.2784066 0.1238493418  0.05639670  0.07827200 -0.02248735
#>  [9,]  0.15018544 -0.2968971 0.0004578713  0.02187781  0.09125045  0.06376733
#> [10,]  0.03595889 -0.3100850 0.0627741871  0.19737472  0.17359402  0.22767745
#>                V24        V25        V26         V27        V28        V29
#>  [1,]  0.227918100 0.11439571 0.25431987  0.21374375 0.17337403 0.18780407
#>  [2,]  0.216907652 0.12328610 0.17028744  0.16728673 0.20459291 0.24817836
#>  [3,]  0.244590106 0.12419172 0.12455211  0.19714875 0.13722837 0.05719605
#>  [4,]  0.147318205 0.22872389 0.28728101  0.27552354 0.31083995 0.22287710
#>  [5,] -0.003074572 0.01658735 0.11856164  0.20707786 0.08560614 0.02258899
#>  [6,]  0.153730924 0.16724642 0.01457292  0.11583712 0.16010809 0.03854640
#>  [7,]  0.163079767 0.23776962 0.27657191  0.11235196 0.23462522 0.24206960
#>  [8,]  0.042749020 0.11512435 0.05665856 -0.04152578 0.02853605 0.03615060
#>  [9,]  0.142505151 0.21143992 0.14914974  0.21034414 0.16112741 0.13818693
#> [10,]  0.223204368 0.17319338 0.12824827  0.29855647 0.20785045 0.17414705
#>               V3         V30          V31         V32         V33          V34
#>  [1,] -0.3529526  0.02983561  0.147697506  0.02998327  0.09443141  0.040401036
#>  [2,] -0.2641354  0.12732688  0.015071489 -0.06366849 -0.01939905  0.008085699
#>  [3,] -0.2323145 -0.01598335  0.067733177 -0.05014459 -0.03643599  0.049774972
#>  [4,] -0.2996893  0.07848819  0.061118242  0.09949790  0.05992810  0.077799300
#>  [5,] -0.2449341  0.03208117  0.001191169  0.10767575  0.02455054 -0.033216764
#>  [6,] -0.2206986  0.06598279 -0.097089648 -0.12604162 -0.16426651 -0.043251120
#>  [7,] -0.3291560  0.04846543  0.013237257 -0.01420296  0.01771764 -0.013459670
#>  [8,] -0.3104421 -0.05228064  0.071392380 -0.09740898 -0.03995440  0.034246496
#>  [9,] -0.3200180  0.18401037  0.069505719 -0.09954361  0.03458831 -0.124444602
#> [10,] -0.1954451  0.26413060  0.115339869  0.12582096 -0.05738562 -0.069425181
#>               V35         V36         V37          V38          V39         V4
#>  [1,] -0.05084291 -0.03363498  0.02224592 -0.070833814 -0.020192572 -0.3399613
#>  [2,]  0.02840517  0.04566356  0.00293913  0.049727791 -0.112016377 -0.1979790
#>  [3,] -0.17212383 -0.10086157 -0.14890431 -0.135837101 -0.025762658 -0.2920691
#>  [4,]  0.04509134 -0.03359204  0.02173085 -0.025335462  0.096851399 -0.2686102
#>  [5,]  0.06434060 -0.07373631  0.07212093  0.082418377 -0.008515241 -0.2322901
#>  [6,] -0.07598784 -0.20592240 -0.20282943 -0.156898255 -0.142615740 -0.1967896
#>  [7,] -0.03454937 -0.04424184 -0.05789986  0.008857425 -0.005909335 -0.2165744
#>  [8,]  0.03931586 -0.06163351 -0.12937871 -0.048209902 -0.052133602 -0.2418491
#>  [9,] -0.10684858 -0.11705502 -0.11898069 -0.177333067 -0.206273012 -0.2663514
#> [10,] -0.05737733  0.11631452  0.07810933  0.001533604 -0.013406357 -0.3310200
#>                V40          V41          V42          V43         V44
#>  [1,] -0.145771748 -0.013468559 -0.036081914 -0.054296093 -0.12049554
#>  [2,] -0.047692636 -0.188315094 -0.064175419 -0.074464712 -0.13739432
#>  [3,] -0.006608859  0.002689136  0.049368613  0.005160982 -0.11840003
#>  [4,] -0.083879403 -0.074171624 -0.002325581 -0.019873086 -0.08731230
#>  [5,]  0.003227720  0.013797423 -0.037245791 -0.111818561 -0.02645301
#>  [6,] -0.110676327 -0.117504826 -0.233995412 -0.055434855 -0.16467351
#>  [7,] -0.085858296 -0.071203919 -0.133387899 -0.118853457 -0.06287815
#>  [8,] -0.053384972 -0.016008802 -0.115222487 -0.267874377 -0.18647653
#>  [9,] -0.109310197 -0.092782246 -0.095459974 -0.147530219 -0.02450547
#> [10,]  0.041267755 -0.156729073 -0.058520910 -0.047646646 -0.15856328
#>                V45         V46         V47        V48        V49         V5
#>  [1,] -0.050654907 -0.28021197 -0.28633872 -0.1533535 -0.3114328 -0.3334822
#>  [2,] -0.126308610 -0.04721577 -0.25038942 -0.3169320 -0.3022193 -0.2807252
#>  [3,] -0.157506244 -0.22242391 -0.08701510 -0.2430907 -0.2764349 -0.1197113
#>  [4,] -0.162823629 -0.12610349 -0.15963994 -0.2397813 -0.2969451 -0.2235657
#>  [5,] -0.017401361 -0.20659904 -0.15103137 -0.1207005 -0.2043784 -0.2514501
#>  [6,] -0.221734359 -0.18983330 -0.17071478 -0.2668501 -0.2374247 -0.1703599
#>  [7,] -0.008076105 -0.04367432 -0.16065871 -0.1886389 -0.2924743 -0.2898531
#>  [8,] -0.276997692 -0.12464974 -0.20609837 -0.2643095 -0.2518591 -0.2547452
#>  [9,] -0.079630080 -0.22472872 -0.20084202 -0.1208372 -0.3299689 -0.2715982
#> [10,] -0.045586184 -0.08987640 -0.08747035 -0.1058929 -0.1560450 -0.2122292
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2676700 -0.3051461 -0.4088860 -0.3410263 -0.3515689 -0.2802561
#>  [2,] -0.4200734 -0.2989343 -0.3138421 -0.4236195 -0.3325410 -0.4006953
#>  [3,] -0.3598884 -0.1930463 -0.3300312 -0.2576753 -0.3155073 -0.2937003
#>  [4,] -0.2484863 -0.3937019 -0.2658688 -0.2795411 -0.3877185 -0.2912617
#>  [5,] -0.2659514 -0.2287351 -0.1904077 -0.2642214 -0.3817297 -0.3419369
#>  [6,] -0.4084295 -0.3553329 -0.3397423 -0.4120038 -0.4576195 -0.2925085
#>  [7,] -0.2802964 -0.3978283 -0.3023571 -0.3571856 -0.2842481 -0.3476373
#>  [8,] -0.2813683 -0.3681860 -0.2633910 -0.2592363 -0.2935363 -0.2745843
#>  [9,] -0.3204072 -0.3898761 -0.3730229 -0.3646852 -0.2459667 -0.2913101
#> [10,] -0.3970687 -0.2770349 -0.3259755 -0.3162564 -0.4001237 -0.4135674
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.4111668 -0.4406397 -0.3694537 -0.4115474 -0.14541759 -0.3235113
#>  [2,] -0.4182590 -0.4402854 -0.4040196 -0.3560721 -0.23424999 -0.3431959
#>  [3,] -0.3404912 -0.2253356 -0.3265020 -0.3266522 -0.09971798 -0.3556755
#>  [4,] -0.4524034 -0.3793870 -0.3754468 -0.3970754 -0.11084102 -0.3284138
#>  [5,] -0.2115589 -0.2255028 -0.1965255 -0.2776286 -0.22439778 -0.3125654
#>  [6,] -0.3841076 -0.4309523 -0.4553744 -0.3267310 -0.24959499 -0.4212644
#>  [7,] -0.2766838 -0.3169695 -0.3471282 -0.3507312 -0.10257462 -0.4096708
#>  [8,] -0.2866118 -0.2603847 -0.3063273 -0.3756105 -0.17115815 -0.2896680
#>  [9,] -0.2989237 -0.4286602 -0.3133180 -0.2734603 -0.18240529 -0.2939030
#> [10,] -0.2722569 -0.2456579 -0.3338625 -0.2502877 -0.28270512 -0.3888737
#>               V7          V8           V9
#>  [1,] -0.1894096 -0.24110619 -0.137789462
#>  [2,] -0.2382925 -0.05498431 -0.110259036
#>  [3,] -0.1490451 -0.09574160 -0.001042491
#>  [4,] -0.1510438 -0.19792265 -0.150410635
#>  [5,] -0.1356950 -0.20586971 -0.059979073
#>  [6,] -0.2594204 -0.15193468 -0.076458713
#>  [7,] -0.1668342 -0.11212313 -0.030423082
#>  [8,] -0.2054261 -0.15193588 -0.094252133
#>  [9,] -0.1651689 -0.15975632 -0.190006439
#> [10,] -0.1387158 -0.21586776 -0.119492351
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]         [,5]        [,6]
#> [1,] -0.08001158  0.11836018  0.09196595  0.07410502 -0.033763998  0.03145498
#> [2,] -0.07995460 -0.07925142 -0.12478953 -0.06202158  0.006909721 -0.07915525
#>             [,7]        [,8]        [,9]        [,10]
#> [1,]  0.05006889 -0.03597784 0.009760207  0.092333259
#> [2,] -0.07521793  0.13563965 0.051537788 -0.004650279
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  5.988555e-05  0.0004470125  0.0005606144  0.0005065192  0.0004651312
#>  [2,] -9.427547e-05 -0.0008692752 -0.0012038852 -0.0009803280 -0.0008536577
#>  [3,] -1.408005e-04 -0.0014254723 -0.0019758517 -0.0016689813 -0.0014271167
#>  [4,] -4.057083e-05 -0.0003534650 -0.0005047789 -0.0003944962 -0.0003474970
#>  [5,]  7.127251e-05  0.0006744452  0.0008734937  0.0008021357  0.0007040469
#>  [6,] -6.251535e-05 -0.0007916825 -0.0011332678 -0.0010310029 -0.0008569606
#>  [7,] -3.044354e-05 -0.0003481564 -0.0005105065 -0.0004002165 -0.0003398101
#>  [8,]  1.638182e-04  0.0018048049  0.0024648294  0.0022234613  0.0018634424
#>  [9,]  1.208294e-04  0.0010272928  0.0013677594  0.0012136000  0.0010770573
#> [10,] -2.117835e-05 -0.0002291246 -0.0003421962 -0.0002629154 -0.0002258880
#>                 V14           V15           V16           V17           V18
#>  [1,]  4.085844e-04  3.808736e-04  3.819826e-04  3.628880e-04  3.875489e-04
#>  [2,] -4.714168e-04 -3.842439e-04 -2.826070e-04 -3.115620e-04 -3.796457e-04
#>  [3,] -8.157283e-04 -6.255492e-04 -4.658216e-04 -4.906348e-04 -6.627240e-04
#>  [4,] -1.714102e-04 -1.480637e-04 -9.358659e-05 -1.110474e-04 -1.495861e-04
#>  [5,]  5.274684e-04  4.615357e-04  4.566147e-04  4.482089e-04  5.208322e-04
#>  [6,] -4.445683e-04 -2.878315e-04 -1.919604e-04 -2.363478e-04 -3.986494e-04
#>  [7,] -1.296693e-04 -9.380719e-05 -3.419029e-05 -5.905469e-05 -1.053018e-04
#>  [8,]  1.135260e-03  8.391695e-04  6.776986e-04  6.932913e-04  9.928796e-04
#>  [9,]  7.872269e-04  6.483715e-04  5.757060e-04  5.691569e-04  6.696078e-04
#> [10,] -7.690152e-05 -4.730259e-05  6.829929e-06 -1.413704e-05 -6.089275e-05
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0005661892  6.790111e-05  0.0007167548  0.0007517094  0.0007073437
#>  [2,] -0.0006745532 -1.044389e-04 -0.0009714299 -0.0011084938 -0.0009676873
#>  [3,] -0.0011861245 -1.612283e-04 -0.0016283454 -0.0018476343 -0.0016578579
#>  [4,] -0.0002647399 -4.663290e-05 -0.0003830882 -0.0004394571 -0.0003530571
#>  [5,]  0.0007510197  8.729530e-05  0.0009296414  0.0009942679  0.0009616982
#>  [6,] -0.0006787384 -8.256866e-05 -0.0008123689 -0.0009246347 -0.0009188573
#>  [7,] -0.0002050328 -3.626683e-05 -0.0003132927 -0.0003795508 -0.0003045894
#>  [8,]  0.0016557412  2.014319e-04  0.0020272785  0.0022241126  0.0021955392
#>  [9,]  0.0011152591  1.340997e-04  0.0014406906  0.0015697676  0.0015075683
#> [10,] -0.0001339001 -2.495995e-05 -0.0001990607 -0.0002407239 -0.0001982198
#>                 V23           V24           V25           V26           V27
#>  [1,]  5.623177e-04  5.527278e-04  0.0006002362  0.0006237941  6.368907e-04
#>  [2,] -4.317645e-04 -4.175577e-04 -0.0007142897 -0.0007118171 -6.473345e-04
#>  [3,] -8.271086e-04 -7.776760e-04 -0.0011987707 -0.0012484622 -1.161156e-03
#>  [4,] -6.766068e-05 -5.610610e-05 -0.0002089362 -0.0002070653 -1.584086e-04
#>  [5,]  7.242657e-04  7.380114e-04  0.0008550255  0.0008956290  9.205406e-04
#>  [6,] -5.162778e-04 -5.081051e-04 -0.0007352672 -0.0007640384 -7.449960e-04
#>  [7,] -5.994046e-06  1.323334e-06 -0.0001639870 -0.0001591368 -1.044504e-04
#>  [8,]  1.401535e-03  1.344190e-03  0.0017287565  0.0017925752  1.777819e-03
#>  [9,]  1.085661e-03  1.028696e-03  0.0011917927  0.0012201402  1.204893e-03
#> [10,] -3.394028e-06  2.912185e-06 -0.0001088483 -0.0001043868 -6.565113e-05
#>                 V28           V29            V3           V30           V31
#>  [1,]  6.135608e-04  5.139628e-04  7.070203e-05  3.711632e-04  0.0001452468
#>  [2,] -5.109900e-04 -2.922120e-04 -1.112852e-04  2.667026e-05  0.0005852404
#>  [3,] -8.718108e-04 -5.084428e-04 -1.716533e-04  6.353558e-06  0.0008950046
#>  [4,] -8.716214e-05 -1.876457e-05 -4.771684e-05  9.880265e-05  0.0003411386
#>  [5,]  8.535236e-04  6.960751e-04  9.106064e-05  4.399070e-04  0.0001072409
#>  [6,] -5.388160e-04 -3.621632e-04 -8.884584e-05 -3.103486e-05  0.0004092160
#>  [7,] -1.248599e-05  6.017381e-05 -3.741585e-05  1.878632e-04  0.0004465503
#>  [8,]  1.457685e-03  1.069169e-03  2.147898e-04  3.537524e-04 -0.0006621181
#>  [9,]  1.066421e-03  8.214599e-04  1.466222e-04  4.591266e-04 -0.0001466221
#> [10,] -4.866602e-06  3.580195e-05 -2.630757e-05  1.172621e-04  0.0002817386
#>                 V32           V33           V34           V35           V36
#>  [1,]  1.958275e-04  0.0001247605  5.182332e-05 -2.399373e-05 -0.0001633887
#>  [2,]  3.903303e-04  0.0005171681  7.208669e-04  1.029963e-03  0.0012947709
#>  [3,]  5.433556e-04  0.0006976045  1.050574e-03  1.537245e-03  0.0018746859
#>  [4,]  2.381697e-04  0.0002822483  3.583348e-04  4.753856e-04  0.0005927629
#>  [5,]  2.148456e-04  0.0001301413 -1.758501e-05 -1.886722e-04 -0.0003189492
#>  [6,]  1.763934e-04  0.0002416283  4.838265e-04  7.944688e-04  0.0009269954
#>  [7,]  3.264936e-04  0.0003526060  4.460997e-04  5.952376e-04  0.0006760140
#>  [8,] -2.521484e-04 -0.0004432761 -9.114847e-04 -1.464237e-03 -0.0018605872
#>  [9,]  2.350145e-05 -0.0001308854 -3.472785e-04 -6.027349e-04 -0.0009159410
#> [10,]  2.032883e-04  0.0002273897  2.949985e-04  3.865324e-04  0.0004491577
#>                 V37           V38           V39            V4           V40
#>  [1,] -8.532136e-05  8.612844e-05  2.023519e-04  9.664159e-05  1.027606e-04
#>  [2,]  9.932750e-04  6.568460e-04  2.492589e-04 -1.717686e-04  2.589094e-04
#>  [3,]  1.363221e-03  8.044015e-04  1.261513e-04 -2.509590e-04  1.359311e-04
#>  [4,]  4.623459e-04  3.329337e-04  1.524764e-04 -7.307400e-05  1.481019e-04
#>  [5,] -1.399607e-04  5.879436e-05  2.686520e-04  1.171407e-04  2.057929e-04
#>  [6,]  6.136074e-04  3.498397e-04 -6.758529e-05 -1.126226e-04 -1.425490e-04
#>  [7,]  5.115525e-04  3.929658e-04  1.957987e-04 -5.703244e-05  1.571843e-04
#>  [8,] -1.239743e-03 -5.717110e-04  2.250166e-04  2.857859e-04  2.418014e-04
#>  [9,] -6.651480e-04 -1.965140e-04  2.116171e-04  2.058713e-04  6.333457e-05
#> [10,]  3.521313e-04  2.762225e-04  1.380346e-04 -3.946263e-05  1.098217e-04
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0001418120  2.745119e-04  0.0003174891  0.0003135599  0.0004010004
#>  [2,]  0.0002411494 -1.218367e-04 -0.0004365056 -0.0004425015 -0.0007030516
#>  [3,]  0.0001806103 -3.809699e-04 -0.0008173736 -0.0008356026 -0.0012459337
#>  [4,]  0.0001662524 -5.354945e-06 -0.0001694941 -0.0001671094 -0.0002805883
#>  [5,]  0.0001915168  3.904000e-04  0.0004680363  0.0004642475  0.0005844593
#>  [6,] -0.0000405668 -3.352869e-04 -0.0005317635 -0.0005234377 -0.0007298088
#>  [7,]  0.0001925058  1.235397e-05 -0.0001548540 -0.0001586237 -0.0002752481
#>  [8,]  0.0002118699  8.068679e-04  0.0011435251  0.0011281622  0.0015712004
#>  [9,]  0.0001502700  5.126177e-04  0.0007092314  0.0007016269  0.0009585894
#> [10,]  0.0001419091  1.724226e-05 -0.0001093269 -0.0001041612 -0.0001818612
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0003181549  1.936947e-04  1.678939e-04  9.343489e-05  1.146046e-04
#>  [2,] -0.0005049875 -2.800611e-04 -2.846393e-04 -1.554479e-04 -1.384819e-04
#>  [3,] -0.0008677277 -4.891029e-04 -4.764405e-04 -2.643859e-04 -2.205326e-04
#>  [4,] -0.0001951597 -1.007665e-04 -1.098908e-04 -5.907922e-05 -4.631867e-05
#>  [5,]  0.0004325437  2.711567e-04  2.447820e-04  1.366806e-04  1.391978e-04
#>  [6,] -0.0004880246 -3.084267e-04 -2.940524e-04 -1.661823e-04 -1.089386e-04
#>  [7,] -0.0001726405 -8.483628e-05 -1.021000e-04 -5.523436e-05 -2.961711e-05
#>  [8,]  0.0011382134  7.065959e-04  6.468381e-04  3.668799e-04  3.112736e-04
#>  [9,]  0.0007304967  4.369813e-04  3.861669e-04  2.165546e-04  2.261150e-04
#> [10,] -0.0001157637 -6.011352e-05 -7.077207e-05 -3.799704e-05 -1.416579e-05
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.455813e-05  3.024883e-05  2.377601e-05  1.241035e-05  1.577782e-05
#>  [2,] -2.172002e-05 -4.363241e-05 -4.724511e-05 -7.290142e-06 -1.606415e-05
#>  [3,] -3.642204e-05 -7.042546e-05 -7.072994e-05 -1.089446e-05 -2.910749e-05
#>  [4,] -5.613725e-06 -1.609545e-05 -2.114131e-05 -2.067926e-06 -5.613220e-06
#>  [5,]  2.906134e-05  3.859000e-05  3.101654e-05  1.283961e-05  1.991653e-05
#>  [6,] -1.972694e-05 -3.784088e-05 -3.365913e-05 -4.165776e-06 -1.784297e-05
#>  [7,] -1.574483e-06 -1.252835e-05 -1.787040e-05  1.215977e-06 -3.821166e-06
#>  [8,]  5.526646e-05  9.132328e-05  7.736441e-05  1.706645e-05  4.366757e-05
#>  [9,]  4.470215e-05  6.452717e-05  5.328766e-05  1.965808e-05  3.099687e-05
#> [10,] -1.185485e-06 -8.892781e-06 -1.341860e-05  2.001681e-07 -2.473479e-06
#>                 V55           V56          V57           V58           V59
#>  [1,]  8.541369e-06  9.686559e-06 5.010363e-06  1.173023e-05  9.056683e-06
#>  [2,] -1.346361e-06 -1.337308e-05 2.418578e-06 -1.921068e-05 -9.744464e-06
#>  [3,] -2.099895e-06 -2.160527e-05 3.230935e-06 -2.927208e-05 -1.427782e-05
#>  [4,]  6.856737e-07 -5.648952e-06 2.785906e-06 -8.779040e-06 -3.896433e-06
#>  [5,]  7.507425e-06  1.323248e-05 5.153929e-06  1.477216e-05  1.103927e-05
#>  [6,]  6.357436e-07 -1.294981e-05 1.326606e-06 -1.542327e-05 -7.089513e-06
#>  [7,]  3.180245e-06 -4.760271e-06 3.995960e-06 -6.838557e-06 -2.110965e-06
#>  [8,]  4.216702e-06  2.892848e-05 1.964727e-06  3.539895e-05  1.910816e-05
#>  [9,]  1.203574e-05  1.874361e-05 4.908131e-06  2.460729e-05  1.570866e-05
#> [10,]  2.342065e-06 -3.308614e-06 3.429187e-06 -5.561950e-06 -1.710484e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.072544e-04  9.079125e-06  1.253197e-04  1.720899e-04  0.0003757214
#>  [2,] -1.032868e-04 -1.385172e-05 -1.327776e-04 -2.294465e-04 -0.0007470274
#>  [3,] -1.828967e-04 -2.167813e-05 -2.412888e-04 -3.996081e-04 -0.0011960129
#>  [4,] -2.705993e-05 -6.210261e-06 -3.835585e-05 -7.800955e-05 -0.0003090421
#>  [5,]  1.507104e-04  1.223388e-05  1.743424e-04  2.513021e-04  0.0005611403
#>  [6,] -1.180915e-04 -1.183575e-05 -1.505902e-04 -2.140989e-04 -0.0006272030
#>  [7,] -1.586097e-05 -5.010410e-06 -2.744495e-05 -7.186668e-05 -0.0003041637
#>  [8,]  3.088902e-04  2.662312e-05  3.720826e-04  5.546229e-04  0.0014531631
#>  [9,]  1.911265e-04  1.724824e-05  2.322768e-04  3.316592e-04  0.0008467393
#> [10,] -5.635975e-06 -3.654373e-06 -1.602484e-05 -4.085951e-05 -0.0001980708
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]         [,5]        [,6]
#> [1,] -0.01469097 -0.02211055 -0.02003905 -0.02175518 -0.004667507 -0.01111179
#> [2,]  0.01469097  0.02211055  0.02003905  0.02175518  0.004667507  0.01111179
#>             [,7]         [,8]        [,9]       [,10]
#> [1,] -0.02329663  0.003130099 -0.01781334 -0.01917221
#> [2,]  0.02329663 -0.003130099  0.01781334  0.01917221
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4235253 0.6134362 0.5178345 0.6421967 0.2660031 0.2677251 0.6638436
#>  [8] 0.1267698 0.5046878 0.5478507
#> 
#> $B[[2]]
#> [1] -0.06299969 -0.06678482
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  6.787406e-04 -2.251656e-05 -2.414215e-04  1.324651e-04  8.314890e-04
#>  [6] -2.545367e-04  2.717628e-04  9.318670e-04  9.238618e-04  1.881821e-04
#> 
#> $vB[[2]]
#> [1] -0.02501998  0.02501998
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#>  [2,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#>  [3,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#>  [4,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#>  [5,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#>  [6,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#>  [7,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#>  [8,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#>  [9,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [10,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [11,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [12,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [13,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [14,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [15,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [16,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [17,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [18,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [19,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [20,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [21,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [22,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [23,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [24,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [25,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#> [26,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [27,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [28,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [29,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [30,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [31,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [32,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [33,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [34,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [35,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [36,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [37,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [38,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [39,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [40,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#>  [2,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#>  [3,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#>  [4,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#>  [5,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#>  [6,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#>  [7,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#>  [8,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#>  [9,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [10,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [11,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [12,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [13,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [14,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [15,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [16,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [17,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [18,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [19,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [20,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [21,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [22,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [23,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [24,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [25,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#> [26,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [27,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [28,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [29,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [30,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [31,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [32,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [33,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [34,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [35,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [36,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [37,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [38,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [39,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [40,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#>  [2,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#>  [3,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#>  [4,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#>  [5,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#>  [6,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#>  [7,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#>  [8,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#>  [9,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [10,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [11,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [12,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [13,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [14,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [15,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [16,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [17,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [18,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [19,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [20,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [21,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [22,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [23,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [24,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [25,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#> [26,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [27,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [28,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [29,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [30,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [31,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [32,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [33,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [34,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [35,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [36,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [37,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [38,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [39,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [40,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#>  [2,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#>  [3,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#>  [4,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#>  [5,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#>  [6,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#>  [7,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#>  [8,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#>  [9,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [10,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [11,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [12,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [13,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [14,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [15,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [16,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [17,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [18,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [19,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [20,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [21,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [22,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [23,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [24,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [25,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#> [26,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [27,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [28,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [29,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [30,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [31,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [32,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [33,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [34,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [35,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [36,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [37,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [38,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [39,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [40,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#>  [2,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#>  [3,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#>  [4,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#>  [5,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#>  [6,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#>  [7,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#>  [8,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#>  [9,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [10,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [11,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [12,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [13,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [14,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [15,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [16,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [17,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [18,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [19,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [20,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [21,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [22,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [23,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [24,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [25,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#> [26,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [27,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [28,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [29,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [30,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [31,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [32,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [33,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [34,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [35,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [36,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [37,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [38,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [39,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [40,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#>  [2,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#>  [3,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#>  [4,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#>  [5,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#>  [6,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#>  [7,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#>  [8,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#>  [9,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [10,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [11,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [12,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [13,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [14,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [15,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [16,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [17,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [18,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [19,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [20,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [21,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [22,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [23,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [24,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [25,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#> [26,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [27,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [28,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [29,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [30,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [31,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [32,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [33,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [34,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [35,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [36,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [37,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [38,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [39,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [40,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8170931 0.8739345 0.7962224 0.9076303 0.6384429 0.6070421 0.8627658
#>  [2,] 0.6391752 0.6968606 0.5529474 0.7881735 0.5728226 0.2745067 0.6747810
#>  [3,] 0.7907666 0.8375164 0.7580904 0.8905967 0.5825110 0.5454934 0.8397760
#>  [4,] 0.7015071 0.8526759 0.7591877 0.8565211 0.4864209 0.5304234 0.8130606
#>  [5,] 0.7736760 0.8560473 0.7555980 0.8936452 0.5807867 0.4929700 0.8476934
#>  [6,] 0.7969606 0.8853382 0.8048794 0.9136133 0.5889405 0.5563962 0.8619332
#>  [7,] 0.7628038 0.8509170 0.7425732 0.8955587 0.5796213 0.4385892 0.8405750
#>  [8,] 0.6195872 0.6854214 0.6024657 0.7844995 0.5671361 0.2850643 0.6892626
#>  [9,] 0.7446547 0.8435404 0.7429916 0.8603883 0.5781533 0.5487869 0.8072932
#> [10,] 0.5828191 0.7401911 0.6249402 0.7569975 0.4687575 0.4366976 0.7024832
#> [11,] 0.8048895 0.8535689 0.7724370 0.9019866 0.6242725 0.5525266 0.8488480
#> [12,] 0.7463616 0.8390259 0.7325368 0.8751213 0.5535625 0.4607816 0.8020572
#> [13,] 0.8441714 0.9035368 0.7972256 0.9295721 0.6588571 0.5722335 0.8721537
#> [14,] 0.8515830 0.9013656 0.7870941 0.9348464 0.6767591 0.5333651 0.8711433
#> [15,] 0.7902060 0.8684448 0.7149837 0.9066353 0.6286619 0.4325763 0.8400874
#> [16,] 0.7513332 0.8442450 0.7526520 0.8862799 0.5542773 0.4862414 0.8389350
#> [17,] 0.7803825 0.8343741 0.7545984 0.8878572 0.6012674 0.5402013 0.8385808
#> [18,] 0.7351810 0.8509202 0.7518795 0.8626523 0.5870556 0.5127051 0.7924265
#> [19,] 0.7652826 0.8134675 0.7284495 0.8757773 0.5936346 0.4963643 0.8189543
#> [20,] 0.7311195 0.8280915 0.7240707 0.8571798 0.5626098 0.5568248 0.8190391
#> [21,] 0.8376916 0.8842854 0.8184242 0.9140607 0.6390215 0.6364412 0.8680716
#> [22,] 0.7263363 0.8470814 0.7476857 0.8710577 0.5854147 0.5232749 0.8185482
#> [23,] 0.7493346 0.8373180 0.7270207 0.8889947 0.5715586 0.4053145 0.8299486
#> [24,] 0.5962786 0.7433568 0.6211590 0.7765399 0.4803743 0.4404264 0.7371290
#> [25,] 0.5298107 0.7199163 0.4816980 0.7591092 0.5239979 0.2012885 0.6755504
#> [26,] 0.6980958 0.7355738 0.5693205 0.8322067 0.6118561 0.2976959 0.7423157
#> [27,] 0.7693293 0.8377353 0.7655896 0.8813765 0.5807324 0.5056245 0.8133256
#> [28,] 0.8139178 0.8631080 0.7837822 0.9039989 0.6248014 0.5791560 0.8557911
#> [29,] 0.8003777 0.8496641 0.7353831 0.8926848 0.6075719 0.5084552 0.8328369
#> [30,] 0.7652231 0.8597956 0.7331944 0.8958155 0.5726658 0.4442775 0.8340250
#> [31,] 0.7244936 0.8332548 0.7350296 0.8668906 0.5513763 0.4492254 0.7997931
#> [32,] 0.8204803 0.8928275 0.8282838 0.9096396 0.6097571 0.6355236 0.8643497
#> [33,] 0.7567952 0.8193271 0.7268872 0.8702235 0.6083948 0.4854559 0.8080465
#> [34,] 0.8064001 0.8856291 0.7973706 0.9132041 0.5868978 0.5928582 0.8720302
#> [35,] 0.7224188 0.8326386 0.7186564 0.8783579 0.5296567 0.4293416 0.8185704
#> [36,] 0.7645498 0.8190849 0.7344082 0.8777290 0.6335425 0.4875708 0.8040116
#> [37,] 0.6317404 0.8053081 0.7171859 0.8173095 0.4619047 0.4669758 0.7649451
#> [38,] 0.8025498 0.8882591 0.8241087 0.9089151 0.6080236 0.6336350 0.8657237
#> [39,] 0.6758215 0.7311988 0.5895335 0.7959931 0.5642991 0.3402580 0.7059437
#> [40,] 0.8042290 0.8492560 0.7775806 0.8994810 0.5928088 0.5619030 0.8499492
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4889366 0.7786482 0.8840152
#>  [2,] 0.3420499 0.4746214 0.7501110
#>  [3,] 0.4275844 0.7364578 0.8517330
#>  [4,] 0.3920418 0.6787236 0.8188485
#>  [5,] 0.3712000 0.7047369 0.8621414
#>  [6,] 0.4371498 0.7436344 0.8833716
#>  [7,] 0.3408101 0.6771461 0.8601310
#>  [8,] 0.3321349 0.4714672 0.7552744
#>  [9,] 0.4631330 0.6900991 0.8302437
#> [10,] 0.3761694 0.5490577 0.6923645
#> [11,] 0.4730806 0.7437073 0.8730107
#> [12,] 0.4461614 0.6371567 0.8451713
#> [13,] 0.5102382 0.7618974 0.9065577
#> [14,] 0.5025433 0.7491363 0.9127935
#> [15,] 0.3951868 0.6576929 0.8747208
#> [16,] 0.3667843 0.6687539 0.8519040
#> [17,] 0.4423317 0.7348940 0.8561178
#> [18,] 0.4450761 0.6672041 0.8340748
#> [19,] 0.4091538 0.7103365 0.8332553
#> [20,] 0.4381019 0.7053426 0.8056984
#> [21,] 0.5150511 0.8006706 0.8990199
#> [22,] 0.4368145 0.6727318 0.8225107
#> [23,] 0.3130995 0.6570788 0.8618114
#> [24,] 0.3582249 0.5865435 0.6844098
#> [25,] 0.2535127 0.3759965 0.7099967
#> [26,] 0.3491865 0.5391176 0.8014869
#> [27,] 0.4551584 0.6839798 0.8347495
#> [28,] 0.4652477 0.7724768 0.8789535
#> [29,] 0.4483914 0.7311708 0.8689698
#> [30,] 0.4038980 0.6442114 0.8410802
#> [31,] 0.4090252 0.6270074 0.8263050
#> [32,] 0.5292349 0.7862582 0.8861958
#> [33,] 0.4022585 0.6991663 0.8471073
#> [34,] 0.4554979 0.7701835 0.8828230
#> [35,] 0.3228209 0.6383394 0.8256356
#> [36,] 0.4626608 0.6666455 0.8385528
#> [37,] 0.3587147 0.6003593 0.7432889
#> [38,] 0.5000087 0.7763069 0.8757989
#> [39,] 0.4165618 0.5558489 0.7655465
#> [40,] 0.4296014 0.7584852 0.8696240
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5717069 0.4282931
#>  [2,] 0.5527817 0.4472183
#>  [3,] 0.5702224 0.4297776
#>  [4,] 0.5731068 0.4268932
#>  [5,] 0.5730735 0.4269265
#>  [6,] 0.5750324 0.4249676
#>  [7,] 0.5728640 0.4271360
#>  [8,] 0.5556906 0.4443094
#>  [9,] 0.5680691 0.4319309
#> [10,] 0.5589703 0.4410297
#> [11,] 0.5697436 0.4302564
#> [12,] 0.5679694 0.4320306
#> [13,] 0.5721207 0.4278793
#> [14,] 0.5712301 0.4287699
#> [15,] 0.5700185 0.4299815
#> [16,] 0.5732207 0.4267793
#> [17,] 0.5690929 0.4309071
#> [18,] 0.5688956 0.4311044
#> [19,] 0.5672547 0.4327453
#> [20,] 0.5676570 0.4323430
#> [21,] 0.5725254 0.4274746
#> [22,] 0.5696413 0.4303587
#> [23,] 0.5723292 0.4276708
#> [24,] 0.5597850 0.4402150
#> [25,] 0.5546175 0.4453825
#> [26,] 0.5560987 0.4439013
#> [27,] 0.5687822 0.4312178
#> [28,] 0.5711351 0.4288649
#> [29,] 0.5679340 0.4320660
#> [30,] 0.5706586 0.4293414
#> [31,] 0.5689914 0.4310086
#> [32,] 0.5730777 0.4269223
#> [33,] 0.5674056 0.4325944
#> [34,] 0.5743971 0.4256029
#> [35,] 0.5720005 0.4279995
#> [36,] 0.5655387 0.4344613
#> [37,] 0.5688723 0.4311277
#> [38,] 0.5740667 0.4259333
#> [39,] 0.5533659 0.4466341
#> [40,] 0.5717943 0.4282057
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]     [,4]        [,5]        [,6]
#>  [1,] 1.4967758 1.9362036  1.36284950 2.285038  0.56861227  0.43489558
#>  [2,] 0.5717861 0.8323926  0.21258642 1.313951  0.29337680 -0.97187603
#>  [3,] 1.3295525 1.6398641  1.14223849 2.096851  0.33308990  0.18247832
#>  [4,] 0.8544847 1.7557450  1.14823120 1.786691 -0.05432983  0.12184427
#>  [5,] 1.2291857 1.7828412  1.12869531 2.128528  0.32600383 -0.02812195
#>  [6,] 1.3674051 2.0439826  1.41707484 2.358574  0.35958702  0.22654878
#>  [7,] 1.1681134 1.7418113  1.05938602 2.148823  0.32121891 -0.24688964
#>  [8,] 0.4877966 0.7787998  0.41574969 1.292083  0.27017615 -0.91947793
#>  [9,] 1.0703041 1.6848101  1.06157611 1.818518  0.31519723  0.19577047
#> [10,] 0.3343569 1.0469622  0.51057033 1.136288 -0.12513287 -0.25457559
#> [11,] 1.4171390 1.7628716  1.22212325 2.219496  0.50772296  0.21088446
#> [12,] 1.0793008 1.6509978  1.00753176 1.947020  0.21507524 -0.15719634
#> [13,] 1.6895986 2.2371549  1.36904347 2.580135  0.65820526  0.29096957
#> [14,] 1.7470708 2.2124911  1.30749724 2.663635  0.73891750  0.13365902
#> [15,] 1.3261676 1.8872776  0.91971319 2.273227  0.52648062 -0.27134752
#> [16,] 1.1057353 1.6901585  1.11280680 2.053293  0.21796797 -0.05504817
#> [17,] 1.2678965 1.6169504  1.12328982 2.069037  0.41074878  0.16115324
#> [18,] 1.0210704 1.7418365  1.10866179 1.837496  0.35180660  0.05083129
#> [19,] 1.1818631 1.4727007  0.98677005 1.953036  0.37901111 -0.01454291
#> [20,] 1.0003100 1.5721616  0.96474436 1.792061  0.25176063  0.22828540
#> [21,] 1.6411519 2.0336529  1.50570747 2.364255  0.57111966  0.55995083
#> [22,] 0.9761130 1.7118913  1.08630710 1.910343  0.34504166  0.09316699
#> [23,] 1.0950668 1.6384064  0.97955921 2.080513  0.28821316 -0.38336929
#> [24,] 0.3899828 1.0634891  0.49447064 1.245615 -0.07854316 -0.23943157
#> [25,] 0.1193843 0.9440465 -0.07324055 1.147802  0.09606548 -1.37826032
#> [26,] 0.8382464 1.0230887  0.27907918 1.601348  0.45512116 -0.85829417
#> [27,] 1.2045282 1.6414729  1.18357297 2.005531  0.32578071  0.02249883
#> [28,] 1.4756710 1.8413477  1.28784507 2.242469  0.50997838  0.31930961
#> [29,] 1.3886569 1.7319693  1.02210829 2.118463  0.43711706  0.03382413
#> [30,] 1.1815318 1.8135932  1.01089062 2.151571  0.29273603 -0.22381985
#> [31,] 0.9668621 1.6088725  1.02029270 1.873741  0.20623293 -0.20380095
#> [32,] 1.5196050 2.1199541  1.57351247 2.309243  0.44629115  0.55598750
#> [33,] 1.1351888 1.5117955  0.97888653 1.902937  0.44056970 -0.05819299
#> [34,] 1.4267864 2.0468515  1.36994101 2.353401  0.35115556  0.37579363
#> [35,] 0.9564915 1.6044440  0.93780683 1.976971  0.11876609 -0.28453799
#> [36,] 1.1777876 1.5101597  1.01710463 1.971098  0.54744469 -0.04972699
#> [37,] 0.5396903 1.4198063  0.93054527 1.498224 -0.15267718 -0.13228926
#> [38,] 1.4023077 2.0730806  1.54443653 2.300460  0.43901212  0.54784293
#> [39,] 0.7346346 1.0007131  0.36203704 1.361437  0.25862833 -0.66214463
#> [40,] 1.4129387 1.7287774  1.25162251 2.191471  0.37558893  0.24888879
#>            [,7]          [,8]       [,9]     [,10]
#>  [1,] 1.8384544 -4.426084e-02  1.2578060 2.0310152
#>  [2,] 0.7298894 -6.541726e-01 -0.1016018 1.0992044
#>  [3,] 1.6565625 -2.917135e-01  1.0276382 1.7482586
#>  [4,] 1.4700213 -4.387376e-01  0.7479122 1.5085652
#>  [5,] 1.7166231 -5.270723e-01  0.8699575 1.8331905
#>  [6,] 1.8314399 -2.527374e-01  1.0649451 2.0247534
#>  [7,] 1.6625127 -6.596862e-01  0.7406873 1.8163781
#>  [8,] 0.7966741 -6.985451e-01 -0.1142551 1.1269434
#>  [9,] 1.4325171 -1.477359e-01  0.8005827 1.5873555
#> [10,] 0.8591506 -5.058394e-01  0.1968640 0.8111970
#> [11,] 1.7255938 -1.077817e-01  1.0653272 1.9278448
#> [12,] 1.3992021 -2.161926e-01  0.5630444 1.6972202
#> [13,] 1.9201370  4.095867e-02  1.1631102 2.2723111
#> [14,] 1.9111054  1.017310e-02  1.0940113 2.3482311
#> [15,] 1.6588787 -4.255610e-01  0.6530300 1.9433602
#> [16,] 1.6503250 -5.460373e-01  0.7025547 1.7496130
#> [17,] 1.6477064 -2.317043e-01  1.0195963 1.7834131
#> [18,] 1.3396142 -2.205859e-01  0.6955667 1.6147857
#> [19,] 1.5092791 -3.674648e-01  0.8970190 1.6088763
#> [20,] 1.5098508 -2.488691e-01  0.8728703 1.4222979
#> [21,] 1.8840148  6.022251e-02  1.3904909 2.1863812
#> [22,] 1.5065420 -2.541003e-01  0.7205669 1.5334513
#> [23,] 1.5852629 -7.856686e-01  0.6503032 1.8304174
#> [24,] 1.0310994 -5.830768e-01  0.3496944 0.7741123
#> [25,] 0.7333974 -1.079965e+00 -0.5065762 0.8953679
#> [26,] 1.0580395 -6.226169e-01  0.1567906 1.3956134
#> [27,] 1.4717656 -1.798495e-01  0.7721221 1.6196694
#> [28,] 1.7807637 -1.392338e-01  1.2223500 1.9825572
#> [29,] 1.6058679 -2.071722e-01  1.0005706 1.8918802
#> [30,] 1.6144264 -3.892495e-01  0.5936902 1.6662876
#> [31,] 1.3850017 -3.679964e-01  0.5193999 1.5596632
#> [32,] 1.8518973  1.170730e-01  1.3025163 2.0524583
#> [33,] 1.4373665 -3.960634e-01  0.8433310 1.7120913
#> [34,] 1.9190301 -1.784808e-01  1.2093479 2.0194394
#> [35,] 1.5066920 -7.408382e-01  0.5681640 1.5550061
#> [36,] 1.4115581 -1.496354e-01  0.6930519 1.6474993
#> [37,] 1.1799848 -5.809472e-01  0.4069623 1.0631337
#> [38,] 1.8636663  3.476377e-05  1.2442729 1.9532346
#> [39,] 0.8757643 -3.369036e-01  0.2243318 1.1833327
#> [40,] 1.7342029 -2.834776e-01  1.1443928 1.8976382
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] 0.07675502 -0.2120636
#>  [2,] 0.04101041 -0.1709061
#>  [3,] 0.07238377 -0.2103747
#>  [4,] 0.08248870 -0.2120495
#>  [5,] 0.07784498 -0.2165572
#>  [6,] 0.08348093 -0.2189326
#>  [7,] 0.07736543 -0.2161804
#>  [8,] 0.04668365 -0.1770069
#>  [9,] 0.07166137 -0.2023160
#> [10,] 0.05615971 -0.1808243
#> [11,] 0.07300437 -0.2078008
#> [12,] 0.07217501 -0.2013960
#> [13,] 0.07810374 -0.2124052
#> [14,] 0.07604624 -0.2108256
#> [15,] 0.07247069 -0.2094559
#> [16,] 0.07856074 -0.2164430
#> [17,] 0.07165863 -0.2064926
#> [18,] 0.07361028 -0.2037364
#> [19,] 0.06705423 -0.2036050
#> [20,] 0.06990639 -0.2023916
#> [21,] 0.07851458 -0.2136476
#> [22,] 0.07457504 -0.2058127
#> [23,] 0.07641307 -0.2149476
#> [24,] 0.05620336 -0.1840862
#> [25,] 0.04811555 -0.1712298
#> [26,] 0.04642399 -0.1789196
#> [27,] 0.07131521 -0.2055690
#> [28,] 0.07515655 -0.2113273
#> [29,] 0.06983662 -0.2035901
#> [30,] 0.07433656 -0.2102020
#> [31,] 0.07314853 -0.2045889
#> [32,] 0.08116275 -0.2132566
#> [33,] 0.06826757 -0.2030064
#> [34,] 0.08228743 -0.2175267
#> [35,] 0.07633267 -0.2136851
#> [36,] 0.06506424 -0.1986076
#> [37,] 0.07402395 -0.2032275
#> [38,] 0.08242736 -0.2160356
#> [39,] 0.04441288 -0.1698669
#> [40,] 0.07538143 -0.2137943
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5717069  0.5717069
#>  [2,] -0.5527817  0.5527817
#>  [3,]  0.4297776 -0.4297776
#>  [4,]  0.4268932 -0.4268932
#>  [5,]  0.4269265 -0.4269265
#>  [6,]  0.4249676 -0.4249676
#>  [7,]  0.4271360 -0.4271360
#>  [8,] -0.5556906  0.5556906
#>  [9,] -0.5680691  0.5680691
#> [10,]  0.4410297 -0.4410297
#> [11,]  0.4302564 -0.4302564
#> [12,] -0.5679694  0.5679694
#> [13,] -0.5721207  0.5721207
#> [14,] -0.5712301  0.5712301
#> [15,]  0.4299815 -0.4299815
#> [16,]  0.4267793 -0.4267793
#> [17,]  0.4309071 -0.4309071
#> [18,]  0.4311044 -0.4311044
#> [19,]  0.4327453 -0.4327453
#> [20,] -0.5676570  0.5676570
#> [21,] -0.5725254  0.5725254
#> [22,] -0.5696413  0.5696413
#> [23,]  0.4276708 -0.4276708
#> [24,]  0.4402150 -0.4402150
#> [25,]  0.4453825 -0.4453825
#> [26,] -0.5560987  0.5560987
#> [27,]  0.4312178 -0.4312178
#> [28,]  0.4288649 -0.4288649
#> [29,] -0.5679340  0.5679340
#> [30,]  0.4293414 -0.4293414
#> [31,]  0.4310086 -0.4310086
#> [32,] -0.5730777  0.5730777
#> [33,]  0.4325944 -0.4325944
#> [34,]  0.4256029 -0.4256029
#> [35,]  0.4279995 -0.4279995
#> [36,] -0.5655387  0.5655387
#> [37,] -0.5688723  0.5688723
#> [38,] -0.5740667  0.5740667
#> [39,] -0.5533659  0.5533659
#> [40,]  0.4282057 -0.4282057
#> 
#> $L
#> [1] 0.6867148 0.6907592 0.6918057 0.7096258 0.7022092 0.6791416
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5072464