Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifDbnDNN$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13         V14
#>  [1,] -0.3792938 -0.04341409 -0.16945496 -0.140693739 -0.10740965  0.08745554
#>  [2,] -0.1862622 -0.09575402 -0.07347882 -0.199136295 -0.15425278 -0.16973057
#>  [3,] -0.3224576 -0.18224266 -0.05076072  0.021009668 -0.06007984 -0.06782122
#>  [4,] -0.3071847 -0.08785827 -0.14806983 -0.100089546 -0.17585605 -0.05471757
#>  [5,] -0.3399428 -0.12381498 -0.16947312 -0.161591206 -0.23858383 -0.13150878
#>  [6,] -0.2233387 -0.26873766 -0.12398292 -0.166388646 -0.04092749 -0.12704020
#>  [7,] -0.3629408 -0.12762789 -0.12367870 -0.094544397 -0.02121863 -0.07014557
#>  [8,] -0.1852747 -0.17463887 -0.13282708 -0.192289520 -0.15247799  0.00997426
#>  [9,] -0.3883406 -0.06147157 -0.02431398 -0.131435056 -0.10722369 -0.11128983
#> [10,] -0.3819826 -0.08740885 -0.11618735 -0.004790265  0.01118631 -0.13453164
#>               V15         V16          V17          V18          V19         V2
#>  [1,]  0.07173679  0.02840066  0.070674021  0.115703546  0.145763237 -0.3149817
#>  [2,] -0.12023595 -0.12308503 -0.162481367 -0.175649215 -0.081262062 -0.2085267
#>  [3,] -0.09348803  0.05237028 -0.109818173 -0.089657393 -0.080702896 -0.3463257
#>  [4,] -0.13717510 -0.18015170 -0.126736820 -0.035630077 -0.008427190 -0.3535384
#>  [5,] -0.03208597 -0.05227108 -0.005436191 -0.100446594 -0.096289155 -0.1943241
#>  [6,]  0.02229943 -0.02750422 -0.130674799 -0.057225550 -0.006137024 -0.2532959
#>  [7,] -0.12495922 -0.03657701 -0.001987414  0.103699622 -0.010319036 -0.2966574
#>  [8,] -0.15714202 -0.03417285  0.039879508 -0.128799834 -0.138683780 -0.2670985
#>  [9,] -0.11423610 -0.05452447  0.052658514  0.007732102  0.009325041 -0.3721640
#> [10,] -0.12457730 -0.10551872  0.035423493 -0.071536124  0.023936504 -0.2058775
#>                 V20         V21          V22         V23          V24
#>  [1,]  0.1324519505  0.12679828  0.243041486  0.17337373  0.126665856
#>  [2,] -0.0119736359  0.04307643 -0.002459574  0.08889707 -0.036356042
#>  [3,] -0.0364588945  0.07758520  0.184944991  0.08062686  0.125489150
#>  [4,] -0.1281911334 -0.04223483 -0.135477087  0.06460033  0.010187141
#>  [5,] -0.0125348452  0.02336110 -0.044735093 -0.08663312  0.005681546
#>  [6,] -0.0516851423  0.01439901  0.127269994  0.09602695  0.068713207
#>  [7,]  0.0195698371  0.03011948  0.050890725  0.11199920  0.239973269
#>  [8,] -0.0005923958 -0.07695080 -0.059977447  0.11621464  0.034449130
#>  [9,]  0.1268625894  0.14757112  0.048322125  0.14930005  0.227658524
#> [10,]  0.0778049621  0.11074907  0.097264021  0.14700487  0.113899459
#>               V25          V26         V27         V28         V29         V3
#>  [1,]  0.11672797  0.155697227  0.21797535  0.21800929  0.17923990 -0.3604188
#>  [2,]  0.05402099 -0.027934740  0.03950247  0.05394670  0.06111511 -0.3710828
#>  [3,]  0.14737664  0.161841355  0.25842645  0.14862363  0.16552773 -0.2822595
#>  [4,]  0.05726540 -0.007744182  0.15068063  0.07911567 -0.06497899 -0.2623543
#>  [5,] -0.07309621  0.015848853 -0.04221544 -0.06760201  0.08332231 -0.3012840
#>  [6,]  0.09892447 -0.013444470  0.13359207 -0.04156061  0.06147334 -0.2165275
#>  [7,]  0.17030144  0.142047312  0.23424955  0.24613786  0.18548132 -0.2250442
#>  [8,]  0.10357378  0.099486286  0.02808412  0.11898911  0.09345643 -0.1929733
#>  [9,]  0.25264070  0.281676887  0.19404247  0.25802501  0.01473603 -0.2696400
#> [10,]  0.11226650  0.229307246  0.08617280  0.14502280  0.20098934 -0.2545792
#>                V30         V31           V32         V33           V34
#>  [1,]  0.031018623  0.11784739  0.0383189723  0.02261347  0.1110543007
#>  [2,] -0.097724478 -0.10196463 -0.0926406837 -0.03330469 -0.0132284648
#>  [3,]  0.135969700  0.05466377 -0.0393842830 -0.05379347 -0.0104453792
#>  [4,] -0.020940252 -0.05622180 -0.0621016015 -0.09121725 -0.1966808928
#>  [5,] -0.016155840 -0.08558108 -0.1116393572 -0.08231416 -0.0662876197
#>  [6,] -0.145975457 -0.04734850  0.0084862658 -0.06135737 -0.1261109785
#>  [7,]  0.105964644  0.02448242  0.0005516484  0.04567258 -0.0002169891
#>  [8,]  0.008209559 -0.03806963 -0.0275359354 -0.09416524 -0.0492473753
#>  [9,]  0.104440236 -0.05372045  0.0199287414  0.02814491  0.0219979434
#> [10,]  0.150457998  0.03587806  0.0620107564  0.11190697  0.1326478379
#>               V35         V36          V37          V38          V39         V4
#>  [1,]  0.04388236 -0.02811679  0.136766185 -0.051861743  0.002201866 -0.2260410
#>  [2,] -0.10131622 -0.04581595  0.040987823 -0.127019137 -0.195600302 -0.3333384
#>  [3,]  0.01197409  0.04490828  0.034651639  0.054979397 -0.098613173 -0.1719704
#>  [4,] -0.12436795 -0.06285546 -0.076437454 -0.020261355 -0.192869339 -0.3055355
#>  [5,] -0.05660281 -0.07713151 -0.001748311 -0.108070088 -0.023011242 -0.2976806
#>  [6,] -0.09851752 -0.10918087 -0.135665434 -0.043143744 -0.071107176 -0.2972618
#>  [7,] -0.09084580 -0.02595419 -0.153256820 -0.064016953 -0.087020100 -0.2975940
#>  [8,] -0.08842865 -0.03496090 -0.159750484 -0.018108051 -0.175175865 -0.2027806
#>  [9,] -0.10741278 -0.01110439 -0.035214248 -0.099928645  0.042011610 -0.2960038
#> [10,]  0.19434475  0.11711973  0.044698756 -0.002399173  0.043266181 -0.2837431
#>               V40          V41         V42         V43          V44         V45
#>  [1,] -0.03665994 -0.083299652 -0.10349043 -0.08076848 -0.141993513 -0.18679658
#>  [2,] -0.10914269 -0.175577768 -0.17900305 -0.10544190 -0.084289071 -0.23608942
#>  [3,] -0.04534472 -0.116158266 -0.05972865 -0.10303664 -0.162435820 -0.05497163
#>  [4,] -0.17747529 -0.114472275 -0.06151078 -0.06352403 -0.131449630 -0.12035791
#>  [5,] -0.08928554 -0.222507680 -0.26298959 -0.19526736 -0.104730204 -0.15534153
#>  [6,] -0.09548281 -0.084179421 -0.15464537 -0.16389317 -0.135915712 -0.20301787
#>  [7,]  0.03841468  0.002731438 -0.01305788 -0.03239268 -0.128809038 -0.11383932
#>  [8,] -0.06795567 -0.068831855 -0.10140390 -0.16627253 -0.040987640 -0.07341697
#>  [9,] -0.14562026 -0.053841085 -0.15134232 -0.11535676 -0.101320121 -0.03654688
#> [10,] -0.06220598  0.044163509 -0.05754690 -0.11425006 -0.007526648  0.01936431
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.12708718 -0.2124867 -0.1380688 -0.2113449 -0.2524306 -0.2800951
#>  [2,] -0.23678064 -0.1440089 -0.1609431 -0.2047911 -0.2084923 -0.3490151
#>  [3,] -0.07585917 -0.2198420 -0.1565808 -0.2509676 -0.2837587 -0.2732789
#>  [4,] -0.11017407 -0.1675614 -0.2937334 -0.2112276 -0.2652680 -0.2498576
#>  [5,] -0.10801877 -0.1740855 -0.2529929 -0.3423845 -0.2135392 -0.2497928
#>  [6,] -0.15995514 -0.1090244 -0.2541414 -0.3705647 -0.1564242 -0.2642104
#>  [7,] -0.24389438 -0.1379805 -0.2696024 -0.3065400 -0.2243115 -0.2541367
#>  [8,] -0.18317234 -0.2221995 -0.2793710 -0.3438788 -0.3001242 -0.3163311
#>  [9,] -0.14780933 -0.2267679 -0.2538969 -0.3069733 -0.2491351 -0.4283873
#> [10,] -0.15199279 -0.2058113 -0.2666160 -0.2950282 -0.2069808 -0.4078500
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.4311941 -0.2916195 -0.3221289 -0.4334823 -0.4387069 -0.3870777
#>  [2,] -0.3021666 -0.3719457 -0.3749475 -0.3730432 -0.2093376 -0.3850144
#>  [3,] -0.3581513 -0.3631785 -0.4134974 -0.3784254 -0.3551595 -0.3798469
#>  [4,] -0.2574418 -0.3577537 -0.4112627 -0.4060004 -0.3271749 -0.2525868
#>  [5,] -0.2808971 -0.1798491 -0.2131306 -0.2438310 -0.1810863 -0.2609149
#>  [6,] -0.3825878 -0.2800990 -0.2595345 -0.2342735 -0.3295516 -0.2769768
#>  [7,] -0.2740396 -0.3777983 -0.2989418 -0.2829919 -0.2672913 -0.2451732
#>  [8,] -0.1877450 -0.2527808 -0.2104546 -0.2901712 -0.3589009 -0.3543018
#>  [9,] -0.3832098 -0.3101732 -0.4455438 -0.2654223 -0.3355511 -0.4076036
#> [10,] -0.3893619 -0.3572293 -0.4061290 -0.4041465 -0.4100362 -0.3810754
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.3941307 -0.3144098 -0.4581015 -0.23824903 -0.3524291 -0.14975764
#>  [2,] -0.2251151 -0.2529272 -0.2298605 -0.18228677 -0.2814352 -0.17038732
#>  [3,] -0.3403794 -0.3959988 -0.4249751 -0.18593020 -0.3714957 -0.08549124
#>  [4,] -0.2889364 -0.2787641 -0.3980710 -0.29586679 -0.4159670 -0.17092352
#>  [5,] -0.1971202 -0.3007612 -0.1906847 -0.15823931 -0.2084232 -0.12137214
#>  [6,] -0.2864868 -0.3037468 -0.2476182 -0.17750391 -0.3345005 -0.32954847
#>  [7,] -0.3823752 -0.2651790 -0.2334744 -0.09997563 -0.2726360 -0.18686985
#>  [8,] -0.3615813 -0.3842254 -0.2893269 -0.21730572 -0.2286240 -0.17273474
#>  [9,] -0.3355339 -0.4673277 -0.3501305 -0.28865088 -0.4755804 -0.14462056
#> [10,] -0.3060628 -0.4326230 -0.3250564 -0.26915988 -0.3607097 -0.28677537
#>                V8          V9
#>  [1,] -0.14526201 -0.08465045
#>  [2,] -0.24384718 -0.21232529
#>  [3,] -0.21495522 -0.06094094
#>  [4,] -0.23328689 -0.04769013
#>  [5,] -0.09283834 -0.08847659
#>  [6,] -0.20536445 -0.16500239
#>  [7,] -0.22354006 -0.18693959
#>  [8,] -0.13146329 -0.20305901
#>  [9,] -0.16745539 -0.11796726
#> [10,] -0.23515564 -0.22134044
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.03162700 -0.0489862  0.09684455 -0.07456507 -0.00470097  0.03667283
#> [2,] -0.07740242  0.0699793 -0.01410379  0.03167211  0.01320004 -0.03431249
#>            [,7]       [,8]        [,9]        [,10]
#> [1,] 0.11606810 0.03624423  0.10936897 -0.001910688
#> [2,] 0.05935286 0.07512574 -0.09359606 -0.036302527
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12          V13
#>  [1,]  1.569809e-04  1.100089e-03  0.0012020038  1.328641e-03 0.0014627400
#>  [2,]  1.879508e-04  1.442170e-03  0.0016761118  1.698661e-03 0.0016180102
#>  [3,]  8.199227e-05  6.305451e-04  0.0006158025  9.222026e-04 0.0012320509
#>  [4,]  1.844796e-04  1.399759e-03  0.0016332038  1.662697e-03 0.0015790317
#>  [5,]  3.874500e-05  2.447223e-04  0.0002229987  3.119073e-04 0.0004520884
#>  [6,]  4.201642e-06 -6.979763e-05 -0.0002394115 -1.492504e-05 0.0003764982
#>  [7,]  1.640705e-04  1.194395e-03  0.0012749106  1.510105e-03 0.0017530957
#>  [8,]  1.708904e-04  1.229314e-03  0.0013444650  1.496920e-03 0.0016533436
#>  [9,] -3.327343e-05 -1.449322e-04 -0.0002782430  1.590167e-04 0.0005408853
#> [10,]  1.741382e-04  1.243727e-03  0.0013599377  1.494351e-03 0.0016591608
#>                V14          V15          V16         V17         V18
#>  [1,] 0.0014364187 0.0015992777 0.0019537372 0.002121544 0.002216599
#>  [2,] 0.0014768050 0.0016559574 0.0020518328 0.002235878 0.002282733
#>  [3,] 0.0013746048 0.0016211335 0.0020805137 0.002359741 0.002622066
#>  [4,] 0.0014276841 0.0015804966 0.0019449399 0.002114255 0.002137288
#>  [5,] 0.0005667197 0.0006687795 0.0008837461 0.001017882 0.001094687
#>  [6,] 0.0007228563 0.0009150147 0.0012548945 0.001481538 0.001678494
#>  [7,] 0.0018239578 0.0020732177 0.0025863975 0.002875086 0.003070139
#>  [8,] 0.0017062409 0.0019333376 0.0024384012 0.002718795 0.002850350
#>  [9,] 0.0007344060 0.0009397531 0.0012923952 0.001534788 0.001890875
#> [10,] 0.0016938055 0.0019124203 0.0023397951 0.002556700 0.002667590
#>               V19            V2         V20         V21          V22
#>  [1,] 0.002543018  2.191520e-04 0.002748089 0.002897131 0.0032195565
#>  [2,] 0.002839096  2.554333e-04 0.003195111 0.003462127 0.0040979989
#>  [3,] 0.002612192  1.101186e-04 0.002626001 0.002677374 0.0027680056
#>  [4,] 0.002686401  2.532459e-04 0.003029902 0.003281382 0.0038855787
#>  [5,] 0.001116303  4.372411e-05 0.001165824 0.001212759 0.0012960531
#>  [6,] 0.001437276 -1.610109e-05 0.001388171 0.001366258 0.0012164390
#>  [7,] 0.003350249  2.240308e-04 0.003551119 0.003704537 0.0040244631
#>  [8,] 0.003224589  2.245372e-04 0.003502775 0.003717605 0.0041589479
#>  [9,] 0.001472601 -5.108974e-05 0.001246704 0.001164578 0.0009768214
#> [10,] 0.003050047  2.439209e-04 0.003276267 0.003436176 0.0037858678
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.003528440 0.003579493 0.003527108 0.003520342 0.003437644 0.003260029
#>  [2,] 0.004299183 0.004133666 0.003980847 0.003712905 0.003489794 0.003012346
#>  [3,] 0.003166233 0.003333220 0.003360282 0.003618329 0.003517122 0.003380007
#>  [4,] 0.004073569 0.003938333 0.003805419 0.003545449 0.003364978 0.002933543
#>  [5,] 0.001528966 0.001606423 0.001637765 0.001730065 0.001683160 0.001585259
#>  [6,] 0.001659940 0.001941904 0.002108894 0.002508249 0.002491913 0.002481017
#>  [7,] 0.004429192 0.004512484 0.004469130 0.004572423 0.004431531 0.004162795
#>  [8,] 0.004543202 0.004562963 0.004516700 0.004508195 0.004339872 0.003963686
#>  [9,] 0.001366682 0.001627410 0.001719051 0.002204338 0.002180139 0.002287553
#> [10,] 0.004099751 0.004138583 0.004060957 0.004054595 0.003924897 0.003645447
#>               V29            V3         V30         V31         V32         V33
#>  [1,] 0.003002012  2.423263e-04 0.002907774 0.002737710 0.002391934 0.002176160
#>  [2,] 0.002460814  2.793681e-04 0.002530960 0.002456187 0.002148447 0.001717100
#>  [3,] 0.003328543  1.296039e-04 0.002995411 0.002786182 0.002386738 0.002444722
#>  [4,] 0.002387254  2.762453e-04 0.002439719 0.002368025 0.002086949 0.001628539
#>  [5,] 0.001526118  5.462128e-05 0.001412901 0.001309377 0.001090485 0.001074092
#>  [6,] 0.002608999 -4.894926e-06 0.002254087 0.002027323 0.001664398 0.001878009
#>  [7,] 0.003901232  2.506684e-04 0.003691761 0.003452503 0.002994353 0.002858266
#>  [8,] 0.003560204  2.534893e-04 0.003430155 0.003232674 0.002798656 0.002531117
#>  [9,] 0.002619139 -4.555630e-05 0.002146591 0.001943553 0.001591170 0.001973530
#> [10,] 0.003284108  2.692833e-04 0.003167396 0.002951838 0.002558999 0.002333177
#>               V34          V35          V36          V37          V38
#>  [1,] 0.001941220 0.0018157531 0.0018454296 0.0017362542 0.0017971652
#>  [2,] 0.001163844 0.0008631634 0.0008653639 0.0010036275 0.0013940274
#>  [3,] 0.002455871 0.0024602077 0.0026661115 0.0024314458 0.0020446116
#>  [4,] 0.001090909 0.0007898041 0.0007848737 0.0009328357 0.0013260115
#>  [5,] 0.001054512 0.0010619816 0.0010952945 0.0009608927 0.0008241326
#>  [6,] 0.002132038 0.0023225849 0.0024361867 0.0020378780 0.0014270104
#>  [7,] 0.002679072 0.0025687746 0.0026677306 0.0024835258 0.0023591097
#>  [8,] 0.002212628 0.0020527447 0.0020897173 0.0019756862 0.0020112520
#>  [9,] 0.002281899 0.0024452293 0.0028191315 0.0024715853 0.0017229816
#> [10,] 0.002059751 0.0018819936 0.0019179252 0.0018504100 0.0019449588
#>                V39            V4          V40          V41          V42
#>  [1,] 0.0018752226  2.746715e-04 0.0018276603 0.0018561128 0.0017831555
#>  [2,] 0.0018405630  3.389306e-04 0.0017262127 0.0019353392 0.0019676181
#>  [3,] 0.0017580700  1.247323e-04 0.0018215994 0.0015002033 0.0012229488
#>  [4,] 0.0017634955  3.337915e-04 0.0016419077 0.0018680546 0.0019187495
#>  [5,] 0.0007127252  5.763852e-05 0.0006983168 0.0006139260 0.0004990983
#>  [6,] 0.0008708412 -2.761955e-05 0.0009310449 0.0005571539 0.0002474571
#>  [7,] 0.0022969601  2.795532e-04 0.0022822442 0.0021575131 0.0019684160
#>  [8,] 0.0021081904  2.941240e-04 0.0020387549 0.0020452709 0.0019193402
#>  [9,] 0.0011454080 -1.048603e-04 0.0013228409 0.0006749870 0.0002518599
#> [10,] 0.0020367122  3.138740e-04 0.0019739979 0.0019898632 0.0019121016
#>                 V43           V44           V45           V46           V47
#>  [1,]  0.0015098725  0.0012904622  0.0012466842  0.0010637517  0.0008419516
#>  [2,]  0.0018040748  0.0016731040  0.0017956766  0.0014449094  0.0011654815
#>  [3,]  0.0009217949  0.0006615043  0.0003052120  0.0002905070  0.0002912002
#>  [4,]  0.0017605387  0.0016487379  0.0017860172  0.0014472757  0.0011543619
#>  [5,]  0.0004094326  0.0003017746  0.0001735113  0.0001613698  0.0001578671
#>  [6,]  0.0001022664 -0.0001054659 -0.0005168271 -0.0003622060 -0.0002295514
#>  [7,]  0.0016252591  0.0013445804  0.0011507566  0.0009767597  0.0007975674
#>  [8,]  0.0016725414  0.0014460502  0.0013546086  0.0011263378  0.0009187305
#>  [9,] -0.0000169504 -0.0002791556 -0.0008828035 -0.0006821540 -0.0004259415
#> [10,]  0.0016198318  0.0013894164  0.0013450056  0.0011316232  0.0008985519
#>                 V48           V49           V5           V50           V51
#>  [1,]  5.754910e-04  3.125428e-04 3.994835e-04  1.289895e-04  1.023774e-04
#>  [2,]  8.207179e-04  4.509317e-04 4.719329e-04  1.534546e-04  1.331940e-04
#>  [3,]  1.756553e-04  9.310535e-05 2.867950e-04  8.068887e-05  4.307111e-05
#>  [4,]  8.126321e-04  4.464268e-04 4.658154e-04  1.504254e-04  1.314677e-04
#>  [5,]  9.642016e-05  4.603157e-05 1.185951e-04  4.192121e-05  2.187895e-05
#>  [6,] -1.957693e-04 -1.242413e-04 5.893918e-05  2.266165e-05 -1.572186e-05
#>  [7,]  5.365109e-04  2.911132e-04 4.586645e-04  1.396256e-04  1.014313e-04
#>  [8,]  6.277777e-04  3.377694e-04 4.554705e-04  1.483297e-04  1.117393e-04
#>  [9,] -3.355403e-04 -1.890532e-04 4.611947e-05 -3.260662e-06 -4.008517e-05
#> [10,]  6.194226e-04  3.396706e-04 4.703885e-04  1.405516e-04  1.119747e-04
#>                 V52          V53           V54           V55          V56
#>  [1,]  8.282597e-05 6.109969e-05  6.569687e-05  5.744611e-05 3.911142e-05
#>  [2,]  1.041160e-04 6.580023e-05  7.861642e-05  6.071736e-05 4.075388e-05
#>  [3,]  3.535174e-05 4.435874e-05  3.989421e-05  3.599257e-05 3.239343e-05
#>  [4,]  1.018273e-04 6.406034e-05  7.711768e-05  5.961969e-05 4.050375e-05
#>  [5,]  2.020894e-05 2.289690e-05  1.961720e-05  1.854683e-05 1.376740e-05
#>  [6,] -5.505648e-06 2.148369e-05  8.204581e-06  1.342208e-05 1.436451e-05
#>  [7,]  8.243443e-05 6.964363e-05  7.141900e-05  6.205834e-05 4.697335e-05
#>  [8,]  9.040780e-05 7.065272e-05  7.494411e-05  6.279379e-05 4.562026e-05
#>  [9,] -3.298385e-05 9.867126e-06 -2.926755e-06 -7.708612e-07 1.319741e-05
#> [10,]  9.173660e-05 6.896208e-05  7.330748e-05  6.305433e-05 4.420368e-05
#>                V57           V58           V59           V6          V60
#>  [1,] 3.500673e-05  4.370076e-05  5.075726e-05 0.0005256867 3.847216e-05
#>  [2,] 2.910285e-05  5.073665e-05  5.982903e-05 0.0005367542 4.184921e-05
#>  [3,] 3.283109e-05  2.211470e-05  3.129895e-05 0.0005164158 2.855176e-05
#>  [4,] 2.941256e-05  5.011890e-05  5.857840e-05 0.0005271337 4.072652e-05
#>  [5,] 1.437855e-05  1.095728e-05  1.548365e-05 0.0001959973 1.252173e-05
#>  [6,] 2.222362e-05 -8.402028e-08  6.148199e-06 0.0002411011 1.031460e-05
#>  [7,] 4.357767e-05  4.504776e-05  5.507974e-05 0.0006626172 4.363163e-05
#>  [8,] 3.996314e-05  4.684316e-05  5.744891e-05 0.0006057372 4.314542e-05
#>  [9,] 1.802056e-05 -8.736918e-06 -1.962097e-07 0.0003225134 8.386690e-06
#> [10,] 3.970186e-05  4.768442e-05  5.618597e-05 0.0006038611 4.244967e-05
#>                 V7           V8            V9
#>  [1,] 0.0006102206 0.0006546726  8.504358e-04
#>  [2,] 0.0005516122 0.0006840496  1.024900e-03
#>  [3,] 0.0006480886 0.0005917964  5.474383e-04
#>  [4,] 0.0005394817 0.0006658542  1.006359e-03
#>  [5,] 0.0002440514 0.0002085304  2.045362e-04
#>  [6,] 0.0003748241 0.0002260121  4.015182e-05
#>  [7,] 0.0007822352 0.0007970549  9.499667e-04
#>  [8,] 0.0006876591 0.0007245909  9.379154e-04
#>  [9,] 0.0004634263 0.0003235228 -5.953530e-06
#> [10,] 0.0006947192 0.0007385703  9.632253e-04
#> 
#> $vW[[2]]
#>             [,1]         [,2]        [,3]         [,4]         [,5]
#> [1,] -0.05347053 -0.001143645 -0.04088406 -0.004952066  0.003802441
#> [2,]  0.05347053  0.001143645  0.04088406  0.004952066 -0.003802441
#>              [,6]        [,7]         [,8]        [,9]       [,10]
#> [1,] -0.008313308 -0.04019546 -0.009859046 -0.04594556 -0.05019719
#> [2,]  0.008313308  0.04019546  0.009859046  0.04594556  0.05019719
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.74254750  0.29422928  0.61032470  0.23113949 -0.03322913  0.15574382
#>  [7]  0.37662156  0.32819267  0.62171642  0.82988850
#> 
#> $B[[2]]
#> [1]  0.04048188 -0.11075223
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.005027673 0.004648555 0.005381776 0.004423197 0.002418911 0.003778404
#>  [7] 0.006519712 0.006057543 0.003807672 0.005661717
#> 
#> $vB[[2]]
#> [1] -0.06968896  0.06968896
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>  [2,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#>  [3,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#>  [4,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#>  [5,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#>  [6,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#>  [7,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#>  [8,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#>  [9,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [10,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [11,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [12,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [13,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [14,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#> [15,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [16,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [17,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [18,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [19,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [20,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [21,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [22,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [23,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [24,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [25,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [26,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [27,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [28,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [29,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [30,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [31,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [32,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [33,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [34,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [35,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [36,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [37,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [38,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [39,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [40,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>  [2,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#>  [3,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#>  [4,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#>  [5,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#>  [6,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#>  [7,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#>  [8,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#>  [9,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [10,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [11,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [12,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [13,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [14,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#> [15,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [16,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [17,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [18,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [19,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [20,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [21,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [22,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [23,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [24,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [25,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [26,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [27,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [28,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [29,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [30,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [31,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [32,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [33,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [34,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [35,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [36,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [37,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [38,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [39,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [40,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>  [2,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#>  [3,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#>  [4,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#>  [5,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#>  [6,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#>  [7,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#>  [8,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#>  [9,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [10,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [11,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [12,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [13,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [14,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#> [15,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [16,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [17,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [18,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [19,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [20,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [21,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [22,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [23,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [24,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [25,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [26,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [27,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [28,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [29,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [30,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [31,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [32,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [33,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [34,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [35,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [36,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [37,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [38,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [39,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [40,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>  [2,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#>  [3,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#>  [4,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#>  [5,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#>  [6,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#>  [7,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#>  [8,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#>  [9,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [10,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [11,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [12,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [13,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [14,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#> [15,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [16,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [17,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [18,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [19,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [20,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [21,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [22,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [23,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [24,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [25,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [26,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [27,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [28,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [29,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [30,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [31,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [32,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [33,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [34,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [35,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [36,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [37,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [38,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [39,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [40,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>  [2,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#>  [3,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#>  [4,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#>  [5,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#>  [6,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#>  [7,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#>  [8,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#>  [9,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [10,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [11,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [12,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [13,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [14,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#> [15,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [16,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [17,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [18,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [19,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [20,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [21,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [22,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [23,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [24,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [25,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [26,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [27,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [28,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [29,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [30,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [31,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [32,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [33,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [34,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [35,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [36,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [37,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [38,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [39,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [40,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#>  [2,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#>  [3,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#>  [4,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#>  [5,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#>  [6,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#>  [7,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#>  [8,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#>  [9,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [10,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [11,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [12,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [13,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [14,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#> [15,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [16,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [17,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [18,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [19,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [20,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [21,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [22,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [23,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [24,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [25,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [26,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [27,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [28,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [29,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [30,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [31,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [32,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [33,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [34,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [35,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [36,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [37,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [38,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [39,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [40,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7664718 0.2970463 0.6443452 0.3057418 0.2471334 0.3398637 0.6288831
#>  [2,] 0.8945842 0.3293105 0.7619009 0.3072440 0.2206912 0.4053505 0.7625783
#>  [3,] 0.6936432 0.1837454 0.5681331 0.1952398 0.1289738 0.2009773 0.5216071
#>  [4,] 0.8271136 0.4371945 0.7726009 0.4484425 0.3197670 0.4572842 0.7506598
#>  [5,] 0.6989005 0.2239999 0.5690752 0.2297442 0.1739616 0.2622813 0.5381278
#>  [6,] 0.8373360 0.3312224 0.7316364 0.2868278 0.2328257 0.3142101 0.6621174
#>  [7,] 0.8429678 0.2536625 0.7289485 0.2506096 0.1648555 0.2931105 0.6995152
#>  [8,] 0.8395960 0.3742701 0.7419025 0.3592249 0.2897253 0.4024512 0.7047367
#>  [9,] 0.7944961 0.2212410 0.5788114 0.2163096 0.1996373 0.2828384 0.5504156
#> [10,] 0.8655894 0.3928207 0.7965240 0.3919861 0.2595909 0.4330216 0.7811174
#> [11,] 0.8798027 0.4147378 0.8169706 0.4090025 0.2645388 0.4570436 0.8083315
#> [12,] 0.8753380 0.2983240 0.7045296 0.2736960 0.2147599 0.3786688 0.7056364
#> [13,] 0.8739673 0.3347919 0.6940275 0.2946834 0.2706385 0.3917431 0.6887421
#> [14,] 0.8638579 0.3216252 0.7565205 0.3036822 0.2245200 0.3580022 0.7632481
#> [15,] 0.7692907 0.2943036 0.6716988 0.2629812 0.2144355 0.2673784 0.6069619
#> [16,] 0.7957870 0.2239943 0.7001679 0.2292290 0.1531001 0.2730530 0.6505233
#> [17,] 0.8506965 0.2898798 0.7442719 0.2715152 0.1865338 0.3281600 0.7173109
#> [18,] 0.8761206 0.3969390 0.7509347 0.3550081 0.2530649 0.4371507 0.7225848
#> [19,] 0.9061786 0.3450280 0.8023536 0.3441828 0.2329465 0.4309754 0.8035867
#> [20,] 0.8182560 0.1760395 0.6203365 0.1975054 0.1224562 0.2390516 0.6300427
#> [21,] 0.8977580 0.3140017 0.7463958 0.2859584 0.2496077 0.3862953 0.7404757
#> [22,] 0.8713809 0.4173228 0.8103658 0.4170506 0.2819143 0.4527040 0.7833861
#> [23,] 0.8308740 0.2751225 0.6892103 0.2302387 0.2066139 0.2867018 0.6533990
#> [24,] 0.8639621 0.4452363 0.7995801 0.4316810 0.3140204 0.4514933 0.7819523
#> [25,] 0.7634432 0.2100333 0.6572665 0.2311189 0.1456272 0.2390757 0.6256082
#> [26,] 0.8634571 0.3277094 0.7070308 0.3334147 0.2502514 0.3677520 0.6981764
#> [27,] 0.8584933 0.3596609 0.7782429 0.3272677 0.2497180 0.3540724 0.7015420
#> [28,] 0.8030055 0.2283321 0.7130199 0.2415050 0.1503168 0.2361702 0.6665328
#> [29,] 0.7760621 0.2058360 0.6605364 0.2300975 0.1489511 0.2385794 0.6113619
#> [30,] 0.8245435 0.3974459 0.7257768 0.3526946 0.3282425 0.3757424 0.6778245
#> [31,] 0.8685259 0.4144141 0.8053523 0.3856937 0.3020402 0.4241160 0.7753274
#> [32,] 0.8857906 0.2950539 0.7359037 0.2603279 0.1874885 0.3561904 0.7464692
#> [33,] 0.8608598 0.2995911 0.6948341 0.2741432 0.2193029 0.3625107 0.6855250
#> [34,] 0.8565016 0.2916137 0.7493523 0.2820858 0.1942519 0.3662438 0.7182899
#> [35,] 0.8875862 0.3075250 0.7479253 0.2990254 0.2284183 0.4062448 0.7365544
#> [36,] 0.8094561 0.2791305 0.6868026 0.2361760 0.1772368 0.2869324 0.6214965
#> [37,] 0.9075996 0.4508776 0.8369727 0.4169443 0.3015867 0.4704546 0.8100835
#> [38,] 0.8082898 0.2409711 0.5821657 0.2299377 0.1903105 0.3080533 0.5684737
#> [39,] 0.8563789 0.3903647 0.7860489 0.3936853 0.2671153 0.4190235 0.7691026
#> [40,] 0.8594293 0.2606757 0.6858340 0.2470179 0.1938974 0.3361921 0.6748045
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4163879 0.6629582 0.7375907
#>  [2,] 0.4500138 0.8167879 0.8558919
#>  [3,] 0.2895791 0.5368197 0.7281126
#>  [4,] 0.5503125 0.7942794 0.8243592
#>  [5,] 0.3244897 0.5704160 0.6753839
#>  [6,] 0.3972072 0.7437124 0.8501239
#>  [7,] 0.3667968 0.7592154 0.8277282
#>  [8,] 0.4825683 0.7561317 0.8042250
#>  [9,] 0.3170358 0.6132546 0.6783036
#> [10,] 0.5064275 0.8221333 0.8362616
#> [11,] 0.5426941 0.8460034 0.8703116
#> [12,] 0.3891698 0.7681339 0.8164692
#> [13,] 0.4282309 0.7539549 0.8130669
#> [14,] 0.4593030 0.7868484 0.8543832
#> [15,] 0.3702162 0.6040782 0.8312131
#> [16,] 0.3656224 0.6891077 0.8351963
#> [17,] 0.3876484 0.7778540 0.8305500
#> [18,] 0.4440987 0.7831573 0.8353997
#> [19,] 0.4933109 0.8483799 0.8660304
#> [20,] 0.3011076 0.7026469 0.7733467
#> [21,] 0.4284028 0.7810080 0.8271229
#> [22,] 0.5255298 0.8369024 0.8490919
#> [23,] 0.3808966 0.7114399 0.8332676
#> [24,] 0.5671617 0.8087193 0.8682008
#> [25,] 0.3296667 0.6599629 0.7704778
#> [26,] 0.4144019 0.7538747 0.8136427
#> [27,] 0.4385826 0.7704933 0.8669213
#> [28,] 0.3537525 0.6907051 0.8471225
#> [29,] 0.3269105 0.6364849 0.7610738
#> [30,] 0.4672305 0.7157679 0.8307167
#> [31,] 0.5344596 0.8042540 0.8885253
#> [32,] 0.3999538 0.8015760 0.8766848
#> [33,] 0.3973707 0.7430590 0.8250618
#> [34,] 0.4027506 0.8005580 0.8319099
#> [35,] 0.4232069 0.8078136 0.8125171
#> [36,] 0.3337802 0.7026897 0.7996450
#> [37,] 0.5540626 0.8603595 0.8994768
#> [38,] 0.3171392 0.6337383 0.7425243
#> [39,] 0.5037665 0.8012639 0.8430177
#> [40,] 0.3711170 0.7442031 0.8059501
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4784745 0.5215255
#>  [2,] 0.4783453 0.5216547
#>  [3,] 0.4832313 0.5167687
#>  [4,] 0.4710511 0.5289489
#>  [5,] 0.4828472 0.5171528
#>  [6,] 0.4777004 0.5222996
#>  [7,] 0.4819029 0.5180971
#>  [8,] 0.4749920 0.5250080
#>  [9,] 0.4833207 0.5166793
#> [10,] 0.4744762 0.5255238
#> [11,] 0.4729761 0.5270239
#> [12,] 0.4803116 0.5196884
#> [13,] 0.4777696 0.5222304
#> [14,] 0.4775141 0.5224859
#> [15,] 0.4766847 0.5233153
#> [16,] 0.4819293 0.5180707
#> [17,] 0.4806741 0.5193259
#> [18,] 0.4751940 0.5248060
#> [19,] 0.4770663 0.5229337
#> [20,] 0.4854541 0.5145459
#> [21,] 0.4788600 0.5211400
#> [22,] 0.4731370 0.5268630
#> [23,] 0.4801951 0.5198049
#> [24,] 0.4701122 0.5298878
#> [25,] 0.4829683 0.5170317
#> [26,] 0.4769784 0.5230216
#> [27,] 0.4758100 0.5241900
#> [28,] 0.4808589 0.5191411
#> [29,] 0.4825692 0.5174308
#> [30,] 0.4729610 0.5270390
#> [31,] 0.4720652 0.5279348
#> [32,] 0.4801682 0.5198318
#> [33,] 0.4793420 0.5206580
#> [34,] 0.4811409 0.5188591
#> [35,] 0.4802516 0.5197484
#> [36,] 0.4814789 0.5185211
#> [37,] 0.4712038 0.5287962
#> [38,] 0.4819096 0.5180904
#> [39,] 0.4737772 0.5262228
#> [40,] 0.4817647 0.5182353
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]       [,6]
#>  [1,] 1.1884952 -0.8614028 0.5942738 -0.8201030 -1.1139600 -0.6639016
#>  [2,] 2.1384464 -0.7113054 1.1631294 -0.8130356 -1.2616432 -0.3832201
#>  [3,] 0.8172073 -1.4911754 0.2742384 -1.4163160 -1.9100624 -1.3801972
#>  [4,] 1.5653075 -0.2525559 1.2230563 -0.2069656 -0.7548430 -0.1712808
#>  [5,] 0.8420674 -1.2425072 0.2780790 -1.2097563 -1.5578065 -1.0341451
#>  [6,] 1.6385384 -0.7026613 1.0029408 -0.9108411 -1.1924240 -0.7805092
#>  [7,] 1.6804776 -1.0791734 0.9892942 -1.0953637 -1.6225354 -0.8803249
#>  [8,] 1.6552255 -0.5139410 1.0558803 -0.5787300 -0.8967185 -0.3952623
#>  [9,] 1.3522431 -1.2584489 0.3178958 -1.2873032 -1.3885628 -0.9304253
#> [10,] 1.8625117 -0.4354710 1.3647094 -0.4389715 -1.0480961 -0.2695336
#> [11,] 1.9905629 -0.3444136 1.4959563 -0.3680906 -1.0225101 -0.1722503
#> [12,] 1.9490044 -0.8552918 0.8689615 -0.9759509 -1.2964687 -0.4952024
#> [13,] 1.9365012 -0.6865907 0.8190162 -0.8727452 -0.9913855 -0.4399912
#> [14,] 1.8477088 -0.7463129 1.1336968 -0.8298245 -1.2395175 -0.5840457
#> [15,] 1.2043105 -0.8745734 0.7158785 -1.0305306 -1.2983938 -1.0079642
#> [16,] 1.3601682 -1.2425394 0.8480974 -1.2126696 -1.7104903 -0.9791877
#> [17,] 1.7400739 -0.8959678 1.0682915 -0.9869489 -1.4726919 -0.7165191
#> [18,] 1.9561954 -0.4182358 1.1036034 -0.5970969 -1.0823322 -0.2527337
#> [19,] 2.2678429 -0.6409668 1.4010696 -0.6447092 -1.1917480 -0.2778728
#> [20,] 1.5045760 -1.5434143 0.4909769 -1.4019595 -1.9693734 -1.1578861
#> [21,] 2.1725577 -0.7814767 1.0794811 -0.9150947 -1.1007057 -0.4629118
#> [22,] 1.9132236 -0.3337734 1.4523888 -0.3348928 -0.9349858 -0.1897513
#> [23,] 1.5918344 -0.9687862 0.7964301 -1.2069638 -1.3454580 -0.9114570
#> [24,] 1.8485953 -0.2199372 1.3836718 -0.2749959 -0.7813901 -0.1946388
#> [25,] 1.1716503 -1.3247244 0.6511363 -1.2020043 -1.7693177 -1.1577537
#> [26,] 1.8443055 -0.7185634 0.8810067 -0.6927810 -1.0972720 -0.5418729
#> [27,] 1.8028321 -0.5768364 1.2554559 -0.7205690 -1.1001167 -0.6011861
#> [28,] 1.4051855 -1.2177534 0.9100963 -1.1444458 -1.7321184 -1.1737921
#> [29,] 1.2428639 -1.3502103 0.6656853 -1.2077609 -1.7428512 -1.1604837
#> [30,] 1.5474385 -0.4161188 0.9733004 -0.6072156 -0.7161446 -0.5076593
#> [31,] 1.8879873 -0.3457472 1.4200886 -0.4654500 -0.8376015 -0.3058992
#> [32,] 2.0484471 -0.8709633 1.0247855 -1.0442649 -1.4664123 -0.5919376
#> [33,] 1.8224494 -0.8492458 0.8228178 -0.9737020 -1.2697333 -0.5644835
#> [34,] 1.7865324 -0.8875595 1.0951611 -0.9341389 -1.4226154 -0.5483651
#> [35,] 2.0663192 -0.8117158 1.0875779 -0.8519431 -1.2172642 -0.3795113
#> [36,] 1.4464797 -0.9487788 0.7852131 -1.1737603 -1.5351816 -0.9103296
#> [37,] 2.2846719 -0.1971256 1.6358740 -0.3353302 -0.8397537 -0.1183194
#> [38,] 1.4389356 -1.1473629 0.3316700 -1.2086629 -1.4479938 -0.8092362
#> [39,] 1.7855344 -0.4457795 1.3012713 -0.4318474 -1.0093078 -0.3267834
#> [40,] 1.8105580 -1.0424593 0.7807142 -1.1145807 -1.4248820 -0.6803100
#>             [,7]        [,8]      [,9]     [,10]
#>  [1,] 0.52742808 -0.33761915 0.6765050 1.0334833
#>  [2,] 1.16686742 -0.20061477 1.4947349 1.7815805
#>  [3,] 0.08648222 -0.89742901 0.1475458 0.9850676
#>  [4,] 1.10213443  0.20193335 1.3509167 1.5461655
#>  [5,] 0.15280776 -0.73321480 0.2835485 0.7326380
#>  [6,] 0.67274434 -0.41711529 1.0653540 1.7355729
#>  [7,] 0.84499024 -0.54598350 1.1483828 1.5696111
#>  [8,] 0.86995670 -0.06975515 1.1315871 1.4129131
#>  [9,] 0.20234993 -0.76742782 0.4610134 0.7459868
#> [10,] 1.27218971  0.02571135 1.5308678 1.6306716
#> [11,] 1.43920521  0.17119318 1.7035928 1.9037162
#> [12,] 0.87428448 -0.45080321 1.1978043 1.4926067
#> [13,] 0.79424499 -0.28907286 1.1198180 1.4700628
#> [14,] 1.17057023 -0.16314896 1.3060317 1.7694010
#> [15,] 0.43455935 -0.53128927 0.4224871 1.5942494
#> [16,] 0.62134010 -0.55104325 0.7959510 1.6229114
#> [17,] 0.93116148 -0.45720797 1.2532039 1.5895299
#> [18,] 0.95731944 -0.22454392 1.2841614 1.6243900
#> [19,] 1.40886421 -0.02675795 1.7219506 1.8663072
#> [20,] 0.53239997 -0.84202913 0.8599343 1.2273061
#> [21,] 1.04844247 -0.28837080 1.2715504 1.5653723
#> [22,] 1.28550906  0.10220796 1.6353587 1.7274968
#> [23,] 0.63401355 -0.48574428 0.9023874 1.6089649
#> [24,] 1.27708010  0.27028023 1.4417102 1.8851437
#> [25,] 0.51342154 -0.70969292 0.6631288 1.2110112
#> [26,] 0.83862914 -0.34579754 1.1193853 1.4738558
#> [27,] 0.85465152 -0.24691652 1.2110986 1.8740075
#> [28,] 0.69254476 -0.60258513 0.8034176 1.7122082
#> [29,] 0.45304048 -0.72219201 0.5601400 1.1585754
#> [30,] 0.74379200 -0.13126602 0.9235650 1.5907153
#> [31,] 1.23864127  0.13805717 1.4130970 2.0757656
#> [32,] 1.07986892 -0.40565776 1.3961736 1.9614039
#> [33,] 0.77928030 -0.41643275 1.0619291 1.5510254
#> [34,] 0.93599462 -0.39401728 1.3897856 1.5992242
#> [35,] 1.02813643 -0.30962251 1.4358656 1.4664492
#> [36,] 0.49590493 -0.69113689 0.8601393 1.3840770
#> [37,] 1.45055273  0.21709923 1.8182787 2.1914246
#> [38,] 0.27562677 -0.76695033 0.5482880 1.0591301
#> [39,] 1.20325105  0.01506619 1.3942128 1.6808552
#> [40,] 0.72999639 -0.52742789 1.0679303 1.4239065
#> 
#> $pre[[3]]
#>              [,1]       [,2]
#>  [1,] -0.02584607 0.06030916
#>  [2,] -0.01540254 0.07127053
#>  [3,] -0.03227086 0.03482919
#>  [4,] -0.02235665 0.09356872
#>  [5,] -0.02867607 0.03996218
#>  [6,] -0.02664217 0.06261543
#>  [7,] -0.01685578 0.05556428
#>  [8,] -0.02305817 0.07705749
#>  [9,] -0.03119922 0.03554276
#> [10,] -0.01591305 0.08627079
#> [11,] -0.01499198 0.09320910
#> [12,] -0.02046291 0.05833148
#> [13,] -0.02508356 0.06389675
#> [14,] -0.01529097 0.07471351
#> [15,] -0.03453811 0.05879077
#> [16,] -0.02040601 0.05190809
#> [17,] -0.01622391 0.06111835
#> [18,] -0.02501619 0.07428923
#> [19,] -0.01140574 0.08039347
#> [20,] -0.02197025 0.03622984
#> [21,] -0.01766407 0.06694644
#> [22,] -0.01737021 0.09018533
#> [23,] -0.02417129 0.05508977
#> [24,] -0.02176182 0.09793215
#> [25,] -0.02177689 0.04637624
#> [26,] -0.02653249 0.06561919
#> [27,] -0.02445427 0.07238149
#> [28,] -0.02215681 0.05444495
#> [29,] -0.02454583 0.04520565
#> [30,] -0.03050123 0.07776032
#> [31,] -0.02009870 0.09175680
#> [32,] -0.01661681 0.06275195
#> [33,] -0.02368241 0.05899673
#> [34,] -0.01444612 0.06102628
#> [35,] -0.01481589 0.06421879
#> [36,] -0.02559017 0.04852805
#> [37,] -0.01851059 0.09680197
#> [38,] -0.03420899 0.03818436
#> [39,] -0.01878864 0.08619901
#> [40,] -0.02111781 0.05185580
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5215255 -0.5215255
#>  [2,]  0.5216547 -0.5216547
#>  [3,]  0.5167687 -0.5167687
#>  [4,] -0.4710511  0.4710511
#>  [5,]  0.5171528 -0.5171528
#>  [6,] -0.4777004  0.4777004
#>  [7,]  0.5180971 -0.5180971
#>  [8,] -0.4749920  0.4749920
#>  [9,] -0.4833207  0.4833207
#> [10,]  0.5255238 -0.5255238
#> [11,]  0.5270239 -0.5270239
#> [12,]  0.5196884 -0.5196884
#> [13,]  0.5222304 -0.5222304
#> [14,] -0.4775141  0.4775141
#> [15,] -0.4766847  0.4766847
#> [16,]  0.5180707 -0.5180707
#> [17,]  0.5193259 -0.5193259
#> [18,] -0.4751940  0.4751940
#> [19,]  0.5229337 -0.5229337
#> [20,]  0.5145459 -0.5145459
#> [21,] -0.4788600  0.4788600
#> [22,]  0.5268630 -0.5268630
#> [23,] -0.4801951  0.4801951
#> [24,] -0.4701122  0.4701122
#> [25,]  0.5170317 -0.5170317
#> [26,] -0.4769784  0.4769784
#> [27,]  0.5241900 -0.5241900
#> [28,]  0.5191411 -0.5191411
#> [29,]  0.5174308 -0.5174308
#> [30,] -0.4729610  0.4729610
#> [31,] -0.4720652  0.4720652
#> [32,]  0.5198318 -0.5198318
#> [33,]  0.5206580 -0.5206580
#> [34,]  0.5188591 -0.5188591
#> [35,]  0.5197484 -0.5197484
#> [36,]  0.5185211 -0.5185211
#> [37,] -0.4712038  0.4712038
#> [38,]  0.5180904 -0.5180904
#> [39,]  0.5262228 -0.5262228
#> [40,]  0.5182353 -0.5182353
#> 
#> $L
#> [1] 0.7169324 0.7956163 0.9156594 0.7959904 0.8904817 0.7035378
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609