Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13          V14
#>  [1,] -0.2354681 -0.03316812 -0.15854755 -0.15191340 -0.06315750 -0.048493338
#>  [2,] -0.3725832 -0.07416500 -0.13765286 -0.01686156 -0.05579474 -0.104628008
#>  [3,] -0.3013370 -0.04770178 -0.01475244 -0.19206546 -0.10453542 -0.043477226
#>  [4,] -0.2835981 -0.04225253 -0.16716703 -0.03760736 -0.07075659  0.022149368
#>  [5,] -0.3107319 -0.03424903 -0.04300147 -0.09533434 -0.09985889 -0.086309857
#>  [6,] -0.2727213 -0.20679795 -0.01415291 -0.13461895 -0.01191280 -0.095033929
#>  [7,] -0.3689240 -0.03261556 -0.14254518 -0.01787454 -0.01247622 -0.071657395
#>  [8,] -0.3044013 -0.13439530 -0.16407984 -0.15156259 -0.13852989 -0.117524379
#>  [9,] -0.3758199 -0.15860568 -0.10712998 -0.06495438  0.02557468  0.001255652
#> [10,] -0.2453470 -0.13804860  0.01016007 -0.03015773 -0.16088466 -0.130778120
#>               V15          V16         V17         V18          V19         V2
#>  [1,] -0.06272252 -0.061536705 -0.15143082 -0.04197376 -0.073975909 -0.2085121
#>  [2,]  0.01694156 -0.103165584 -0.00639957 -0.08902805 -0.053928525 -0.1688150
#>  [3,] -0.19740875 -0.105667610 -0.14461648 -0.16746329 -0.074412124 -0.3027490
#>  [4,]  0.08166378 -0.019105210  0.08847419  0.09972849  0.122133031 -0.2072205
#>  [5,] -0.08858872 -0.154088653 -0.14965288 -0.03471320 -0.008496664 -0.2586447
#>  [6,] -0.02970145 -0.025023720  0.06009795  0.05637096  0.039882727 -0.3695295
#>  [7,]  0.06815696  0.062222845 -0.06454736  0.07163510  0.146232210 -0.2942609
#>  [8,] -0.09807555  0.006018243  0.02868992  0.10203989  0.040152927 -0.2039020
#>  [9,]  0.07979912  0.080433402  0.15413110  0.07556445  0.187675320 -0.3794402
#> [10,] -0.10849603 -0.038482955 -0.03322590  0.06135017  0.001342350 -0.3754050
#>                V20          V21         V22        V23          V24         V25
#>  [1,] -0.122284259  0.004942451 -0.05991243 0.08595707  0.037919331 -0.10068577
#>  [2,] -0.005360478 -0.055079496  0.06118812 0.17063670  0.090993987  0.07647135
#>  [3,] -0.060432711  0.027481802 -0.07889144 0.01350909  0.040707306  0.06815642
#>  [4,]  0.042891934  0.149489781  0.11238074 0.03373758  0.006549108 -0.01337289
#>  [5,] -0.038944033  0.070104139  0.05688906 0.10405014 -0.026533216  0.10558178
#>  [6,]  0.094501506  0.102320996  0.04613380 0.16909555  0.068409886  0.09453660
#>  [7,]  0.181528796  0.095893916  0.25824575 0.11698596  0.096034721  0.09217475
#>  [8,]  0.117834200  0.078636977  0.19579166 0.07508957  0.200127066  0.24790848
#>  [9,]  0.089447810  0.070833397  0.06973618 0.08972970  0.158800909  0.08362312
#> [10,]  0.075065346  0.162618921  0.08226306 0.01745964  0.128292603  0.15977209
#>                V26         V27        V28          V29         V3         V30
#>  [1,] -0.037956800 -0.06489936 0.13469116  0.076357353 -0.3491436 -0.14214783
#>  [2,] -0.007891639  0.13896015 0.03705334  0.158610224 -0.2551503  0.10548404
#>  [3,] -0.079479925 -0.03384830 0.16132134  0.092580842 -0.3231880 -0.01130920
#>  [4,] -0.013983337  0.07332076 0.10058136  0.062863927 -0.3345661 -0.01182850
#>  [5,]  0.167067799  0.16338953 0.13825763  0.030916427 -0.2387959  0.03967808
#>  [6,]  0.058017294  0.19317145 0.10344370 -0.008599771 -0.2557168  0.13080972
#>  [7,]  0.120942781  0.16995913 0.14235668  0.043107917 -0.2570532  0.05533275
#>  [8,]  0.223444917  0.13963993 0.05167498  0.136934117 -0.1793688  0.11313536
#>  [9,]  0.277548944  0.22817240 0.23742744  0.230057585 -0.2087135  0.20836272
#> [10,]  0.245953594  0.16240827 0.11596507  0.153437117 -0.2435989  0.03656529
#>               V31         V32          V33         V34           V35
#>  [1,]  0.03532058  0.01990491 -0.114075010 -0.02737473 -0.1440935228
#>  [2,]  0.01535966  0.05886220 -0.053523075  0.06759589 -0.0635273382
#>  [3,] -0.03690304  0.06655787  0.060558645  0.03831148  0.0002405837
#>  [4,]  0.03968614  0.02617986 -0.083740349 -0.07823831 -0.0055999358
#>  [5,] -0.13033470 -0.04778783 -0.063181034 -0.18302217 -0.1687048338
#>  [6,]  0.12503901  0.05621475 -0.026097996 -0.03885803 -0.0085364294
#>  [7,]  0.02777725 -0.11346166  0.014088062 -0.09938986 -0.0463360293
#>  [8,] -0.03596896 -0.06970239  0.021216286 -0.01253001 -0.0165293625
#>  [9,]  0.19845601  0.13029965  0.037313390 -0.03788111  0.0024811162
#> [10,] -0.07272239  0.04938614 -0.003053461 -0.03266998 -0.0291903950
#>               V36          V37          V38         V39         V4         V40
#>  [1,] -0.09508460 -0.126552663 -0.027372340 -0.03874640 -0.1280756 -0.16932120
#>  [2,]  0.04262115  0.086638464  0.004425686 -0.06666955 -0.2195221 -0.03060080
#>  [3,]  0.05853092 -0.063236210  0.035575164  0.01350953 -0.3381947 -0.05471264
#>  [4,] -0.04938342 -0.091431477 -0.147762379 -0.08523891 -0.2207422 -0.07170620
#>  [5,] -0.04392090 -0.082246558 -0.106757998 -0.17294041 -0.1856544 -0.07059386
#>  [6,] -0.01061578 -0.046993367 -0.068799755 -0.08107543 -0.3139794  0.04081264
#>  [7,] -0.05133012 -0.060451747 -0.086398207 -0.07547964 -0.2292800 -0.14225762
#>  [8,] -0.03194826  0.014284748  0.032001959 -0.03746409 -0.3101746 -0.08122123
#>  [9,] -0.04232678  0.001444934 -0.060322011  0.03787233 -0.2435967  0.00912391
#> [10,] -0.05528058  0.030606771 -0.075714948 -0.04401657 -0.1831017 -0.13940064
#>               V41          V42         V43         V44         V45         V46
#>  [1,] -0.02747467 -0.189932457 -0.20350860 -0.21871067 -0.06514546 -0.15074592
#>  [2,] -0.07434557  0.059577895 -0.10594606  0.01564316 -0.04494269 -0.19694657
#>  [3,] -0.08827911  0.053583371 -0.05107661 -0.05720251 -0.06092393 -0.08275783
#>  [4,] -0.03660445 -0.161417034 -0.02744125 -0.05855497 -0.08431121 -0.15998070
#>  [5,] -0.10306190 -0.186362222 -0.17559602 -0.12916921 -0.13766550 -0.13487937
#>  [6,] -0.06531692 -0.049637338 -0.10188188 -0.10483100 -0.06066036 -0.25640842
#>  [7,] -0.06319741 -0.076227825 -0.11813116 -0.09634411 -0.15163314 -0.20357423
#>  [8,] -0.08726321 -0.123422100 -0.04245954 -0.14566989 -0.20134808 -0.17217707
#>  [9,] -0.01608230 -0.089770990 -0.08594400 -0.08243121 -0.06357492 -0.07715303
#> [10,] -0.12888873 -0.007984275 -0.17604257 -0.20531991 -0.19962443 -0.07287470
#>              V47         V48        V49         V5        V50        V51
#>  [1,] -0.2530741 -0.21109596 -0.1819677 -0.1978589 -0.2427379 -0.1827400
#>  [2,] -0.1958571 -0.08665701 -0.2674377 -0.1865031 -0.3253953 -0.3182271
#>  [3,] -0.2082508 -0.27486014 -0.2002498 -0.2883096 -0.2870171 -0.2629430
#>  [4,] -0.1359203 -0.18408294 -0.3336923 -0.3027845 -0.3382649 -0.4245104
#>  [5,] -0.1930607 -0.25342317 -0.3213896 -0.1776777 -0.4063316 -0.2199423
#>  [6,] -0.2385351 -0.27691285 -0.3731329 -0.2897480 -0.3111466 -0.3636454
#>  [7,] -0.2657001 -0.20380324 -0.3477617 -0.2014465 -0.3571687 -0.2891058
#>  [8,] -0.1782757 -0.28274012 -0.1767005 -0.2054759 -0.2590027 -0.3315619
#>  [9,] -0.1419109 -0.21605365 -0.1991996 -0.1317731 -0.3982432 -0.3749952
#> [10,] -0.1739064 -0.15408110 -0.2845296 -0.1873636 -0.2573750 -0.3788557
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3075358 -0.3652442 -0.3578678 -0.3019675 -0.2687678 -0.3130686
#>  [2,] -0.3097269 -0.2629749 -0.2762346 -0.1985315 -0.2316191 -0.3923685
#>  [3,] -0.3757992 -0.3374732 -0.2161742 -0.3834598 -0.2875029 -0.3406161
#>  [4,] -0.2856605 -0.3516655 -0.3246715 -0.3819258 -0.3128124 -0.3092265
#>  [5,] -0.3417768 -0.2686554 -0.2622356 -0.2334704 -0.3711170 -0.2999964
#>  [6,] -0.2514275 -0.3063371 -0.3674671 -0.3824093 -0.3605972 -0.2827328
#>  [7,] -0.3462305 -0.3693083 -0.2968344 -0.4038811 -0.3845241 -0.3015010
#>  [8,] -0.4011404 -0.4012703 -0.3140412 -0.3154920 -0.3725057 -0.2494474
#>  [9,] -0.3299077 -0.3570668 -0.3792673 -0.3988788 -0.4685256 -0.4548549
#> [10,] -0.2858400 -0.2991650 -0.3978475 -0.3128414 -0.2839487 -0.3712800
#>              V58        V59          V6        V60          V7          V8
#>  [1,] -0.3291537 -0.2562538 -0.15832997 -0.2112384 -0.14313444 -0.26820724
#>  [2,] -0.2791648 -0.3013927 -0.12086597 -0.3745956 -0.22021687 -0.18752928
#>  [3,] -0.3895822 -0.3385410 -0.18120797 -0.2957644 -0.09275848 -0.13029674
#>  [4,] -0.3289698 -0.3597753 -0.24852386 -0.4287997 -0.20736706 -0.15462068
#>  [5,] -0.3299057 -0.3174616 -0.31233520 -0.3110686 -0.24999091 -0.20781153
#>  [6,] -0.3689640 -0.3129023 -0.27404045 -0.4502803 -0.13803031 -0.11602184
#>  [7,] -0.4542984 -0.3282140 -0.17563679 -0.3878300 -0.21653794 -0.11896611
#>  [8,] -0.3276626 -0.4175734 -0.11628119 -0.2960323 -0.13363807 -0.08053843
#>  [9,] -0.3032039 -0.4405383 -0.08072551 -0.4260976 -0.20286843 -0.16291988
#> [10,] -0.2757937 -0.3519752 -0.22876228 -0.2867897 -0.11492567 -0.15648869
#>                V9
#>  [1,] -0.06814936
#>  [2,] -0.08213958
#>  [3,] -0.06910668
#>  [4,] -0.05906037
#>  [5,] -0.22720441
#>  [6,] -0.14159616
#>  [7,] -0.10838396
#>  [8,] -0.20879725
#>  [9,] -0.03039337
#> [10,] -0.01499303
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] -0.05532082 0.02261823 -0.03912682 -0.1180997 -0.01654011 -0.14179718
#> [2,]  0.09842918 0.16183722  0.11739206  0.0226803  0.02545335  0.08575966
#>             [,7]       [,8]        [,9]       [,10]
#> [1,] -0.06715049 -0.1717692 -0.06685641 -0.01149804
#> [2,]  0.01010751  0.1537761  0.08574382  0.16521558
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 1.943827e-05 1.038907e-04 0.0001196243 0.0001507564 0.0001362143
#>  [2,] 5.148580e-05 3.335295e-04 0.0004187250 0.0004955203 0.0007689190
#>  [3,] 4.228301e-05 2.304269e-04 0.0002723638 0.0003223610 0.0004089905
#>  [4,] 5.678488e-05 3.496499e-04 0.0004327155 0.0004970064 0.0007362370
#>  [5,] 2.397632e-05 9.148745e-05 0.0001150282 0.0001517271 0.0005896150
#>  [6,] 7.589776e-05 4.770977e-04 0.0005369231 0.0005790531 0.0006467418
#>  [7,] 3.210264e-05 1.936222e-04 0.0003207994 0.0004325348 0.0008581581
#>  [8,] 1.092266e-04 6.982795e-04 0.0007124093 0.0006953177 0.0005637035
#>  [9,] 3.983774e-05 2.911960e-04 0.0003625776 0.0004163601 0.0005859584
#> [10,] 6.451412e-05 4.165498e-04 0.0004958244 0.0005557956 0.0007327986
#>                V14           V15           V16          V17          V18
#>  [1,] 5.529552e-05 -2.223618e-05 -1.233637e-05 0.0001470310 0.0002964518
#>  [2,] 1.135309e-03  1.359775e-03  1.719043e-03 0.0020175045 0.0020625127
#>  [3,] 4.698463e-04  4.692120e-04  6.072226e-04 0.0008516735 0.0010040177
#>  [4,] 1.053412e-03  1.230838e-03  1.552582e-03 0.0018429585 0.0019204506
#>  [5,] 1.196826e-03  1.555404e-03  1.965003e-03 0.0021908239 0.0020803773
#>  [6,] 6.798366e-04  6.472511e-04  8.239425e-04 0.0011099952 0.0013114944
#>  [7,] 1.467825e-03  1.877862e-03  2.367094e-03 0.0026774382 0.0026212208
#>  [8,] 2.786185e-04 -7.369053e-06 -3.032436e-06 0.0002594986 0.0005983458
#>  [9,] 8.205504e-04  9.554980e-04  1.209613e-03 0.0013944603 0.0014222020
#> [10,] 9.545605e-04  1.069840e-03  1.347324e-03 0.0016277998 0.0017392398
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0002540451 2.637809e-05 0.0002394853 0.0002724221 0.0002628285
#>  [2,] 0.0017747273 7.765210e-05 0.0017344664 0.0018869594 0.0021568863
#>  [3,] 0.0009078806 5.754767e-05 0.0009191421 0.0010086889 0.0010940349
#>  [4,] 0.0016903605 8.162960e-05 0.0016926883 0.0018495837 0.0021003747
#>  [5,] 0.0016774076 3.674624e-05 0.0016729034 0.0018552147 0.0022147908
#>  [6,] 0.0012927575 1.012371e-04 0.0013753697 0.0014755604 0.0015271130
#>  [7,] 0.0021107103 5.796284e-05 0.0019649094 0.0021684354 0.0026547259
#>  [8,] 0.0008840833 1.324159e-04 0.0011439844 0.0012001245 0.0009965201
#>  [9,] 0.0012603263 6.146534e-05 0.0012388430 0.0013470113 0.0015455073
#> [10,] 0.0015741146 9.111865e-05 0.0015881424 0.0017151092 0.0018970371
#>                V23          V24          V25           V26           V27
#>  [1,] 0.0002007840 0.0001510579 0.0001158847 -7.526527e-05 -7.914832e-05
#>  [2,] 0.0024479420 0.0027368203 0.0028610242  2.862892e-03  2.590660e-03
#>  [3,] 0.0011678450 0.0012387249 0.0012479884  1.097277e-03  1.004592e-03
#>  [4,] 0.0023858315 0.0026501138 0.0027506096  2.734823e-03  2.494681e-03
#>  [5,] 0.0027108426 0.0031906445 0.0034577771  3.651278e-03  3.234062e-03
#>  [6,] 0.0015798006 0.0016063796 0.0015824749  1.428985e-03  1.367329e-03
#>  [7,] 0.0031725137 0.0036996153 0.0039303853  4.044807e-03  3.604295e-03
#>  [8,] 0.0008121786 0.0005730712 0.0004155194  1.665879e-04  3.290093e-04
#>  [9,] 0.0017647734 0.0019625309 0.0020254720  2.053076e-03  1.900684e-03
#> [10,] 0.0020902115 0.0022615190 0.0023075418  2.249371e-03  2.074207e-03
#>                 V28          V29           V3          V30          V31
#>  [1,] -4.447455e-05 0.0000100175 3.138485e-05 2.167673e-05 8.554756e-05
#>  [2,]  2.509501e-03 0.0025722240 1.045490e-04 2.372204e-03 2.209956e-03
#>  [3,]  1.010242e-03 0.0010719108 7.252193e-05 1.000630e-03 9.710112e-04
#>  [4,]  2.438227e-03 0.0025120867 1.079800e-04 2.318912e-03 2.147531e-03
#>  [5,]  3.119091e-03 0.0032022715 5.885628e-05 2.992418e-03 2.705178e-03
#>  [6,]  1.376548e-03 0.0013971508 1.228440e-04 1.308748e-03 1.245251e-03
#>  [7,]  3.488089e-03 0.0036486134 9.097260e-05 3.316434e-03 3.063592e-03
#>  [8,]  4.181489e-04 0.0003546736 1.462655e-04 3.972724e-04 4.056392e-04
#>  [9,]  1.852887e-03 0.0018685719 8.031098e-05 1.673485e-03 1.559030e-03
#> [10,]  2.041887e-03 0.0020942743 1.157326e-04 1.940033e-03 1.815026e-03
#>                V32          V33          V34           V35           V36
#>  [1,] 0.0002146065 0.0001510240 3.096874e-05 -4.500973e-05 -0.0001569474
#>  [2,] 0.0016825141 0.0017492222 1.921472e-03  2.010232e-03  0.0021619111
#>  [3,] 0.0008942499 0.0008619438 8.266196e-04  7.847070e-04  0.0007160262
#>  [4,] 0.0016884013 0.0017593218 1.905610e-03  1.977514e-03  0.0020775311
#>  [5,] 0.0019533915 0.0021954562 2.609942e-03  2.806322e-03  0.0029802481
#>  [6,] 0.0011825752 0.0011051164 1.006387e-03  9.322014e-04  0.0008265501
#>  [7,] 0.0021685894 0.0023949094 2.805686e-03  3.058446e-03  0.0034464241
#>  [8,] 0.0008031675 0.0005684706 1.853253e-04 -8.375923e-05 -0.0004950318
#>  [9,] 0.0011826938 0.0012303536 1.339755e-03  1.405737e-03  0.0015710957
#> [10,] 0.0014785141 0.0014869150 1.542700e-03  1.567142e-03  0.0016273717
#>                 V37          V38          V39           V4          V40
#>  [1,] -0.0001329211 2.691838e-05 0.0001685463 2.966307e-05 0.0001948076
#>  [2,]  0.0018304363 1.286779e-03 0.0010379099 9.135617e-05 0.0011304877
#>  [3,]  0.0006030610 5.598333e-04 0.0005960805 6.973200e-05 0.0006289832
#>  [4,]  0.0017599112 1.279540e-03 0.0010629487 9.851469e-05 0.0011317944
#>  [5,]  0.0023769219 1.465324e-03 0.0009501449 3.594090e-05 0.0010872939
#>  [6,]  0.0007549700 7.946621e-04 0.0008923436 1.366864e-04 0.0008844750
#>  [7,]  0.0028742498 1.809826e-03 0.0012636778 4.946629e-05 0.0014172694
#>  [8,] -0.0002976658 3.494653e-04 0.0007725836 1.970404e-04 0.0006514335
#>  [9,]  0.0013676995 9.840518e-04 0.0008205820 7.325032e-05 0.0008549521
#> [10,]  0.0014121206 1.110135e-03 0.0010076722 1.148180e-04 0.0010505639
#>                V41          V42          V43           V44           V45
#>  [1,] 0.0002716721 0.0003396145 2.987535e-04  2.629767e-04  2.193668e-04
#>  [2,] 0.0009013768 0.0005579227 4.354414e-04  3.330071e-04  1.902675e-05
#>  [3,] 0.0006161639 0.0005639647 4.854538e-04  4.082872e-04  2.681427e-04
#>  [4,] 0.0009242037 0.0006344457 5.215114e-04  4.026633e-04  1.029463e-04
#>  [5,] 0.0007829412 0.0002459865 8.800573e-05 -3.168330e-05 -3.563448e-04
#>  [6,] 0.0008897842 0.0008903013 7.997070e-04  7.043039e-04  5.722372e-04
#>  [7,] 0.0009363687 0.0003192014 1.939479e-04  5.034138e-05 -4.789826e-04
#>  [8,] 0.0009346054 0.0012894688 1.213363e-03  1.136076e-03  1.248881e-03
#>  [9,] 0.0006572782 0.0004243712 3.655690e-04  2.947619e-04  4.515969e-05
#> [10,] 0.0009049220 0.0007075329 6.070611e-04  5.040824e-04  2.613343e-04
#>                 V46           V47           V48           V49           V5
#>  [1,]  1.702670e-04  1.636723e-04  1.223463e-04  6.579197e-05 4.117682e-05
#>  [2,]  1.629697e-05  4.893228e-05  1.366002e-05  1.724090e-06 1.968241e-04
#>  [3,]  2.176872e-04  1.944130e-04  1.340299e-04  7.079032e-05 1.187489e-04
#>  [4,]  8.435204e-05  9.039425e-05  4.546233e-05  1.954657e-05 1.979729e-04
#>  [5,] -2.628589e-04 -2.116128e-04 -1.966053e-04 -1.202593e-04 1.345972e-04
#>  [6,]  4.681056e-04  3.706179e-04  2.699890e-04  1.487065e-04 1.980748e-04
#>  [7,] -4.045271e-04 -2.564647e-04 -2.245848e-04 -1.337395e-04 2.006863e-04
#>  [8,]  1.033650e-03  7.569476e-04  5.723109e-04  3.225152e-04 1.995698e-04
#>  [9,]  2.238658e-05  4.154029e-05  2.173915e-05  9.676767e-06 1.592172e-04
#> [10,]  2.112126e-04  1.866279e-04  1.242455e-04  6.595862e-05 2.024959e-04
#>                 V50           V51           V52          V53          V54
#>  [1,]  2.495690e-05  1.924995e-05  8.645043e-06 9.931302e-06 1.677974e-05
#>  [2,]  3.237981e-05  1.658019e-05  1.196276e-05 2.538596e-05 1.989455e-05
#>  [3,]  3.958347e-05  2.699894e-05  1.573827e-05 2.110588e-05 2.436586e-05
#>  [4,]  3.699707e-05  2.136734e-05  1.577438e-05 2.818051e-05 2.331373e-05
#>  [5,]  1.104427e-05 -1.055566e-05 -2.746443e-06 1.696030e-05 7.214739e-06
#>  [6,]  6.521802e-05  4.948209e-05  3.207976e-05 3.419582e-05 3.770237e-05
#>  [7,] -2.073286e-06 -1.250533e-05 -5.257762e-06 2.027755e-05 5.024546e-06
#>  [8,]  1.100683e-04  8.894031e-05  5.828254e-05 4.525499e-05 5.826567e-05
#>  [9,]  2.083322e-05  1.452095e-05  1.125443e-05 2.044951e-05 1.343256e-05
#> [10,]  4.559597e-05  3.093912e-05  2.120311e-05 2.967393e-05 2.722808e-05
#>                V55          V56          V57           V58          V59
#>  [1,] 8.415939e-06 5.164824e-06 2.175713e-07  9.405840e-06 6.969262e-06
#>  [2,] 2.826435e-05 2.164698e-05 3.077197e-05  1.117251e-05 1.628313e-05
#>  [3,] 1.976631e-05 1.424706e-05 1.281962e-05  1.487274e-05 1.397221e-05
#>  [4,] 2.912680e-05 2.228697e-05 2.992694e-05  1.449031e-05 1.819477e-05
#>  [5,] 2.708123e-05 1.789305e-05 3.615547e-05 -6.676870e-07 9.087853e-06
#>  [6,] 2.976154e-05 2.339722e-05 1.850647e-05  2.729133e-05 2.332230e-05
#>  [7,] 2.629525e-05 1.965911e-05 4.174171e-05 -6.399647e-07 1.248998e-05
#>  [8,] 3.511952e-05 2.716208e-05 6.800450e-06  4.653092e-05 3.171650e-05
#>  [9,] 1.987868e-05 1.613516e-05 2.301726e-05  9.097794e-06 1.265091e-05
#> [10,] 2.863247e-05 2.261727e-05 2.603307e-05  1.845517e-05 1.985640e-05
#>                 V6          V60           V7           V8           V9
#>  [1,] 1.111008e-05 4.060966e-06 1.766493e-05 1.173371e-05 4.118775e-05
#>  [2,] 3.832944e-04 1.788818e-05 4.875089e-04 4.644004e-04 3.707392e-04
#>  [3,] 1.556806e-04 1.118228e-05 1.967176e-04 1.833740e-04 1.912275e-04
#>  [4,] 3.628377e-04 1.844434e-05 4.577077e-04 4.397516e-04 3.704695e-04
#>  [5,] 3.884681e-04 1.668139e-05 5.379888e-04 4.464869e-04 2.413375e-04
#>  [6,] 2.239561e-04 1.705648e-05 2.668843e-04 2.950595e-04 3.871593e-04
#>  [7,] 5.283315e-04 1.956673e-05 6.651591e-04 6.054136e-04 3.300785e-04
#>  [8,] 5.955482e-05 1.764020e-05 5.967029e-05 1.386348e-04 4.726156e-04
#>  [9,] 2.986179e-04 1.302345e-05 3.579805e-04 3.641648e-04 3.110845e-04
#> [10,] 3.235708e-04 1.798018e-05 3.993028e-04 4.078701e-04 3.955878e-04
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.0549815  0.09137969  0.06891237  0.08704054  0.06548892  0.09758642
#> [2,] -0.0549815 -0.09137969 -0.06891237 -0.08704054 -0.06548892 -0.09758642
#>            [,7]       [,8]       [,9]       [,10]
#> [1,]  0.0954666  0.1031223  0.1098876  0.09582237
#> [2,] -0.0954666 -0.1031223 -0.1098876 -0.09582237
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.2217057 0.4789989 0.2399501 0.5423944 0.1025130 0.5178345 0.4010328
#>  [8] 0.5313100 0.5758004 0.4665375
#> 
#> $B[[2]]
#> [1] -0.1008609  0.0489291
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0001377936 0.0040980214 0.0017571694 0.0039416186 0.0048966659
#>  [6] 0.0023217004 0.0056312246 0.0007277153 0.0029707801 0.0033643825
#> 
#> $vB[[2]]
#> [1]  0.1221236 -0.1221236
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#>  [2,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#>  [3,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#>  [4,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#>  [5,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>  [6,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#>  [7,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#>  [8,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#>  [9,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [10,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [11,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [12,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [13,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [14,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [15,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [16,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [17,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [18,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [19,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [20,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [21,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [22,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [23,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [24,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [25,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [26,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [27,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [28,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [29,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [30,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [31,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [32,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [33,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [34,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [35,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [36,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [37,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [38,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [39,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [40,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#>  [2,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#>  [3,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#>  [4,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#>  [5,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>  [6,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#>  [7,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#>  [8,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#>  [9,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [10,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [11,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [12,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [13,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [14,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [15,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [16,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [17,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [18,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [19,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [20,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [21,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [22,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [23,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [24,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [25,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [26,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [27,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [28,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [29,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [30,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [31,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [32,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [33,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [34,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [35,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [36,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [37,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [38,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [39,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [40,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#>  [2,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#>  [3,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#>  [4,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#>  [5,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>  [6,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#>  [7,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#>  [8,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#>  [9,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [10,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [11,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [12,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [13,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [14,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [15,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [16,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [17,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [18,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [19,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [20,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [21,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [22,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [23,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [24,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [25,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [26,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [27,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [28,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [29,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [30,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [31,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [32,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [33,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [34,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [35,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [36,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [37,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [38,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [39,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [40,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#>  [2,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#>  [3,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#>  [4,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#>  [5,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>  [6,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#>  [7,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#>  [8,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#>  [9,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [10,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [11,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [12,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [13,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [14,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [15,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [16,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [17,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [18,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [19,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [20,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [21,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [22,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [23,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [24,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [25,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [26,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [27,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [28,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [29,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [30,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [31,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [32,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [33,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [34,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [35,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [36,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [37,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [38,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [39,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [40,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#>  [2,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#>  [3,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#>  [4,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#>  [5,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>  [6,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#>  [7,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#>  [8,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#>  [9,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [10,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [11,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [12,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [13,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [14,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [15,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [16,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [17,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [18,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [19,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [20,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [21,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [22,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [23,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [24,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [25,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [26,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [27,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [28,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [29,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [30,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [31,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [32,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [33,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [34,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [35,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [36,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [37,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [38,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [39,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [40,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#>  [2,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#>  [3,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#>  [4,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#>  [5,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#>  [6,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#>  [7,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#>  [8,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#>  [9,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [10,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [11,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [12,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [13,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [14,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [15,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [16,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [17,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [18,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [19,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [20,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [21,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [22,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [23,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [24,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [25,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [26,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [27,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [28,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [29,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [30,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [31,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [32,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [33,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [34,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [35,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [36,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [37,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [38,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [39,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [40,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2549149 0.5131775 0.2878718 0.6164124 0.2686619 0.5934415 0.6313192
#>  [2,] 0.2565710 0.5430626 0.3234835 0.6264371 0.2788548 0.6292978 0.6541383
#>  [3,] 0.3242475 0.6162031 0.4196210 0.6258690 0.4232620 0.6729121 0.7196031
#>  [4,] 0.2564850 0.6277311 0.4162056 0.5779213 0.3501345 0.6509526 0.6538131
#>  [5,] 0.2870936 0.6192996 0.3566965 0.6843545 0.3940033 0.7134810 0.7615620
#>  [6,] 0.2781154 0.6500697 0.5006708 0.5100147 0.3148434 0.6203836 0.5372318
#>  [7,] 0.3817498 0.7308915 0.5251322 0.6725728 0.5088053 0.7683647 0.7739579
#>  [8,] 0.3524581 0.6377211 0.3942169 0.7238856 0.4884268 0.7588391 0.7996073
#>  [9,] 0.3340112 0.6870126 0.4613981 0.6576582 0.4665103 0.7379774 0.7541611
#> [10,] 0.2069041 0.5369258 0.3077199 0.5802194 0.2356840 0.6080352 0.6271358
#> [11,] 0.3969526 0.6578579 0.5289457 0.5933326 0.3985729 0.6565633 0.6271251
#> [12,] 0.2968889 0.6392373 0.4472305 0.5922657 0.3784619 0.6797774 0.6477905
#> [13,] 0.2632760 0.6100731 0.3842481 0.6196569 0.3262159 0.6799745 0.6885668
#> [14,] 0.2387684 0.5824453 0.3297745 0.5896890 0.3408390 0.6450207 0.7007130
#> [15,] 0.4074456 0.6648412 0.5118091 0.6227331 0.5061580 0.6953914 0.6843977
#> [16,] 0.2750714 0.6205871 0.3994485 0.6368734 0.3475261 0.6848389 0.6943709
#> [17,] 0.3784499 0.6781500 0.4781739 0.6309002 0.4934425 0.7132706 0.7180139
#> [18,] 0.2357182 0.5706930 0.3891832 0.5167265 0.2790335 0.5743024 0.5353183
#> [19,] 0.4019362 0.7123035 0.5087886 0.6444894 0.5237059 0.7436855 0.7395149
#> [20,] 0.3875261 0.7359984 0.5373784 0.6712623 0.5232628 0.7758399 0.7728146
#> [21,] 0.2722188 0.6038272 0.3892158 0.5646068 0.3037228 0.6270473 0.6319888
#> [22,] 0.2858107 0.5829078 0.3891183 0.5862739 0.3176839 0.6345491 0.6332956
#> [23,] 0.2008174 0.5338666 0.3261668 0.5225475 0.2558337 0.5936299 0.5866168
#> [24,] 0.4003451 0.7199193 0.5686756 0.6149190 0.4647546 0.7265413 0.6721643
#> [25,] 0.2810525 0.6104207 0.3765194 0.6553515 0.3746900 0.6964891 0.7364145
#> [26,] 0.3167441 0.6556242 0.3953534 0.6784163 0.4537855 0.7345159 0.7669652
#> [27,] 0.4023663 0.6671095 0.4860278 0.6298468 0.4661929 0.6980394 0.7049690
#> [28,] 0.2274493 0.5703862 0.3739375 0.5243883 0.2542412 0.5632421 0.5437393
#> [29,] 0.2734597 0.4901321 0.3071378 0.6310678 0.3013844 0.6012922 0.6364356
#> [30,] 0.2300022 0.6328153 0.4595216 0.4665931 0.2661088 0.5789730 0.4909748
#> [31,] 0.3967023 0.7326433 0.5434189 0.6675110 0.5284804 0.7702551 0.7607717
#> [32,] 0.2488430 0.6197097 0.4317069 0.5235856 0.2781204 0.6083420 0.5555037
#> [33,] 0.3794987 0.6944070 0.5007841 0.6576086 0.4949596 0.7380800 0.7335234
#> [34,] 0.3495855 0.6611074 0.4978822 0.6073735 0.4395092 0.6807452 0.6677420
#> [35,] 0.2777431 0.6201392 0.3458451 0.6868643 0.3693547 0.7075664 0.7460583
#> [36,] 0.2575317 0.5982017 0.3245849 0.6703519 0.3499998 0.6985672 0.7305196
#> [37,] 0.3171524 0.6612198 0.4666645 0.6160858 0.4238074 0.7032494 0.6896305
#> [38,] 0.4143132 0.7090229 0.5531801 0.6469452 0.4960400 0.7302895 0.6972979
#> [39,] 0.2649279 0.6310799 0.3637684 0.6118008 0.3507130 0.6663521 0.6787275
#> [40,] 0.3158139 0.6900038 0.4808143 0.6329744 0.4066910 0.7377567 0.7034585
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5966596 0.7997406 0.5466029
#>  [2,] 0.6351144 0.7805316 0.5624315
#>  [3,] 0.7357215 0.8354004 0.6891072
#>  [4,] 0.6725264 0.8638089 0.6444920
#>  [5,] 0.7610010 0.8795937 0.6922112
#>  [6,] 0.6535251 0.8145678 0.6003746
#>  [7,] 0.8269519 0.9008069 0.7802940
#>  [8,] 0.8157916 0.8924395 0.7542716
#>  [9,] 0.7812130 0.8952532 0.7372110
#> [10,] 0.5964927 0.8377404 0.5504465
#> [11,] 0.6986163 0.8180926 0.6556484
#> [12,] 0.6856694 0.8497210 0.6376452
#> [13,] 0.7398797 0.8448803 0.6498696
#> [14,] 0.7331797 0.8199409 0.6426031
#> [15,] 0.7477312 0.8366617 0.7174054
#> [16,] 0.7393256 0.8421866 0.6494504
#> [17,] 0.7663405 0.8755839 0.7366051
#> [18,] 0.5682449 0.8249803 0.5451428
#> [19,] 0.7867300 0.8947577 0.7584582
#> [20,] 0.8306094 0.9036675 0.7895310
#> [21,] 0.6826365 0.7934475 0.5984762
#> [22,] 0.6546756 0.8559499 0.6058674
#> [23,] 0.6289858 0.8410016 0.5642906
#> [24,] 0.7622849 0.8793783 0.7182288
#> [25,] 0.7558501 0.8527356 0.6750573
#> [26,] 0.7950058 0.8963719 0.7347654
#> [27,] 0.7393116 0.8272727 0.6941533
#> [28,] 0.5893026 0.8180310 0.5443394
#> [29,] 0.5849596 0.7988281 0.5731761
#> [30,] 0.5960970 0.8267859 0.5578402
#> [31,] 0.8245220 0.8966178 0.7822401
#> [32,] 0.6109799 0.8269433 0.5790754
#> [33,] 0.7779774 0.8867458 0.7374542
#> [34,] 0.7131421 0.8570709 0.6965931
#> [35,] 0.7619705 0.8976649 0.6967023
#> [36,] 0.7384615 0.8891489 0.6763862
#> [37,] 0.7462680 0.8611690 0.6974739
#> [38,] 0.7601563 0.8723884 0.7246588
#> [39,] 0.7247188 0.8488264 0.6421770
#> [40,] 0.7828945 0.9000392 0.7291460
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5468474 0.4531526
#>  [2,] 0.5462513 0.4537487
#>  [3,] 0.5489415 0.4510585
#>  [4,] 0.5483083 0.4516917
#>  [5,] 0.5503616 0.4496384
#>  [6,] 0.5431227 0.4568773
#>  [7,] 0.5508299 0.4491701
#>  [8,] 0.5516150 0.4483850
#>  [9,] 0.5509004 0.4490996
#> [10,] 0.5469235 0.4530765
#> [11,] 0.5454944 0.4545056
#> [12,] 0.5475442 0.4524558
#> [13,] 0.5464692 0.4535308
#> [14,] 0.5471432 0.4528568
#> [15,] 0.5485373 0.4514627
#> [16,] 0.5470906 0.4529094
#> [17,] 0.5499094 0.4500906
#> [18,] 0.5453167 0.4546833
#> [19,] 0.5508183 0.4491817
#> [20,] 0.5509549 0.4490451
#> [21,] 0.5448076 0.4551924
#> [22,] 0.5467933 0.4532067
#> [23,] 0.5449580 0.4550420
#> [24,] 0.5476575 0.4523425
#> [25,] 0.5486458 0.4513542
#> [26,] 0.5509360 0.4490640
#> [27,] 0.5484787 0.4515213
#> [28,] 0.5445745 0.4554255
#> [29,] 0.5476608 0.4523392
#> [30,] 0.5429891 0.4570109
#> [31,] 0.5506169 0.4493831
#> [32,] 0.5447325 0.4552675
#> [33,] 0.5503047 0.4496953
#> [34,] 0.5485267 0.4514733
#> [35,] 0.5498997 0.4501003
#> [36,] 0.5495308 0.4504692
#> [37,] 0.5482784 0.4517216
#> [38,] 0.5490617 0.4509383
#> [39,] 0.5475656 0.4524344
#> [40,] 0.5480940 0.4519060
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]        [,2]         [,3]        [,4]        [,5]      [,6]
#>  [1,] -1.0725689  0.05272224 -0.905742495  0.47434807 -1.00142207 0.3782108
#>  [2,] -1.0638678  0.17267816 -0.737808597  0.51696179 -0.95014948 0.5292055
#>  [3,] -0.7343197  0.47346293 -0.324329280  0.51453499 -0.30939680 0.7213858
#>  [4,] -1.0643185  0.52249550 -0.338369483  0.31424603 -0.61844807 0.6232292
#>  [5,] -0.9095419  0.48657646 -0.589731381  0.77385631 -0.43051542 0.9123508
#>  [6,] -0.9538292  0.61934573  0.002683235  0.04006417 -0.77757196 0.4911765
#>  [7,] -0.4821278  0.99915023  0.100613758  0.71984443  0.03522487 1.1991005
#>  [8,] -0.6082520  0.56548655 -0.429620941  0.96381826 -0.04630093 1.1463254
#>  [9,] -0.6900984  0.78618990 -0.154715531  0.65287588 -0.13415958 1.0354822
#> [10,] -1.3436888  0.14797262 -0.810800615  0.32367421 -1.17648927 0.4390608
#> [11,] -0.4181791  0.65376270  0.115912369  0.37775954 -0.41141502 0.6480163
#> [12,] -0.8621568  0.57205539 -0.211866914  0.37333960 -0.49608192 0.7527491
#> [13,] -1.0290105  0.44761959 -0.471555718  0.48809217 -0.72535036 0.7536548
#> [14,] -1.1594437  0.33281976 -0.709205124  0.36267997 -0.65955741 0.5972230
#> [15,] -0.3745354  0.68494369  0.047245181  0.50116511  0.02463307 0.8254474
#> [16,] -0.9690424  0.49204095 -0.407763425  0.56181951 -0.62993127 0.7760996
#> [17,] -0.4961329  0.74528276 -0.087359959  0.53608044 -0.02623152 0.9113218
#> [18,] -1.1762997  0.28467933 -0.450747065  0.06693081 -0.94926101 0.2994268
#> [19,] -0.3974040  0.90659819  0.035158111  0.59490337  0.09489476 1.0652131
#> [20,] -0.4577233  1.02527285  0.149792874  0.71390003  0.09311850 1.2415854
#> [21,] -0.9833942  0.42143741 -0.450609774  0.25988017 -0.82963252 0.5195704
#> [22,] -0.9158183  0.33472186 -0.451019954  0.34858287 -0.76443598 0.5517825
#> [23,] -1.3811935  0.13567406 -0.725573589  0.09025130 -1.06773705 0.3789919
#> [24,] -0.4040273  0.94406134  0.276449587  0.46803665 -0.14121582 0.9771447
#> [25,] -0.9392469  0.44908102 -0.504348149  0.64264669 -0.51214863 0.8306347
#> [26,] -0.7687753  0.64385432 -0.424864284  0.74650333 -0.18538705 1.0176569
#> [27,] -0.3956151  0.69514052 -0.055903347  0.53155951 -0.13543507 0.8379790
#> [28,] -1.2227701  0.28342695 -0.515361689  0.09763073 -1.07611903 0.2543306
#> [29,] -0.9771398 -0.03947679 -0.813534670  0.53680036 -0.84071432 0.4108521
#> [30,] -1.2082990  0.54431350 -0.162268851 -0.13382701 -1.01445534 0.3185588
#> [31,] -0.4192246  1.00807522  0.174114203  0.69694926  0.11404521 1.2097521
#> [32,] -1.1047923  0.48831641 -0.274890618  0.09441251 -0.95380402 0.4403482
#> [33,] -0.4916764  0.82080393  0.003136537  0.65265534 -0.02016245 1.0360132
#> [34,] -0.6208615  0.66823294 -0.008471146  0.43628524 -0.24315404 0.7571986
#> [35,] -0.9556843  0.49013898 -0.637353342  0.78550020 -0.53498599 0.8835940
#> [36,] -1.0588372  0.39797789 -0.732780513  0.70977700 -0.61903991 0.8404845
#> [37,] -0.7668893  0.66873462 -0.133540219  0.47296733 -0.30716307 0.8628197
#> [38,] -0.3461629  0.89064313  0.213528213  0.60563813 -0.01584027 0.9960921
#> [39,] -1.0205107  0.53685228 -0.559045213  0.45488827 -0.61590652 0.6917319
#> [40,] -0.7730770  0.80013728 -0.076780546  0.54499836 -0.37766177 1.0343414
#>              [,7]      [,8]     [,9]     [,10]
#>  [1,]  0.53788017 0.3915660 1.384674 0.1869543
#>  [2,]  0.63727960 0.5542213 1.268767 0.2510363
#>  [3,]  0.94249373 1.0238481 1.624395 0.7959486
#>  [4,]  0.63584257 0.7196340 1.847293 0.5949145
#>  [5,]  1.16126246 1.1581753 1.988589 0.8104773
#>  [6,]  0.14920326 0.6345706 1.479968 0.4070260
#>  [7,]  1.23079613 1.5641768 2.206223 1.2673805
#>  [8,]  1.38384164 1.4880913 2.115905 1.1215257
#>  [9,]  1.12092972 1.2727493 2.145560 1.0315225
#> [10,]  0.51994877 0.3908722 1.641511 0.2024751
#> [11,]  0.51990308 0.8407174 1.503478 0.6439613
#> [12,]  0.60934116 0.7799505 1.732415 0.5651583
#> [13,]  0.79342731 1.0453433 1.694998 0.6184660
#> [14,]  0.85069518 1.0108153 1.515947 0.5866802
#> [15,]  0.77405620 1.0865484 1.633596 0.9316279
#> [16,]  0.82063385 1.0424662 1.674589 0.6166244
#> [17,]  0.93463097 1.1877617 1.951260 1.0283973
#> [18,]  0.14150881 0.2746940 1.550461 0.1810642
#> [19,]  1.04344858 1.3053263 2.140287 1.1442454
#> [20,]  1.22427274 1.5899523 2.238656 1.3221007
#> [21,]  0.54075818 0.7659147 1.345833 0.3991200
#> [22,]  0.54638105 0.6396556 1.782051 0.4299737
#> [23,]  0.34999679 0.5278683 1.665700 0.2585937
#> [24,]  0.71799041 1.1652474 1.986556 0.9356925
#> [25,]  1.02741515 1.1300609 1.756220 0.7311489
#> [26,]  1.19125373 1.3553676 2.157546 1.0189364
#> [27,]  0.87107349 1.0423934 1.566421 0.8196091
#> [28,]  0.17540551 0.3610830 1.503064 0.1778246
#> [29,]  0.55992672 0.3431670 1.378986 0.2948217
#> [30,] -0.03610468 0.3892287 1.563017 0.2324011
#> [31,]  1.15691516 1.5472904 2.160197 1.2787688
#> [32,]  0.22293338 0.4514332 1.564117 0.3189788
#> [33,]  1.01257295 1.2539183 2.057924 1.0327782
#> [34,]  0.69798989 0.9106936 1.791172 0.8311268
#> [35,]  1.07769900 1.1635134 2.171544 0.8316433
#> [36,]  0.99726065 1.0379874 2.082076 0.7372133
#> [37,]  0.79839265 1.0788062 1.825033 0.8352976
#> [38,]  0.83446341 1.1535366 1.922244 0.9676897
#> [39,]  0.74792990 0.9679907 1.725426 0.5848254
#> [40,]  0.86382189 1.2826144 2.197661 0.9902942
#> 
#> $pre[[3]]
#>            [,1]        [,2]
#>  [1,] 0.1134440 -0.07449679
#>  [2,] 0.1150279 -0.07050785
#>  [3,] 0.1393545 -0.05704033
#>  [4,] 0.1394060 -0.05443164
#>  [5,] 0.1344198 -0.06771219
#>  [6,] 0.1383584 -0.03456196
#>  [7,] 0.1598488 -0.04417563
#>  [8,] 0.1421060 -0.06509205
#>  [9,] 0.1514082 -0.05290118
#> [10,] 0.1174758 -0.07077230
#> [11,] 0.1442219 -0.03826032
#> [12,] 0.1390385 -0.05171477
#> [13,] 0.1286882 -0.05772643
#> [14,] 0.1262343 -0.06290028
#> [15,] 0.1513972 -0.04336552
#> [16,] 0.1306799 -0.05824255
#> [17,] 0.1532241 -0.04708047
#> [18,] 0.1276700 -0.05409544
#> [19,] 0.1596131 -0.04436437
#> [20,] 0.1618302 -0.04269926
#> [21,] 0.1268504 -0.05286232
#> [22,] 0.1294061 -0.05831662
#> [23,] 0.1190228 -0.06129605
#> [24,] 0.1567732 -0.03443714
#> [25,] 0.1317460 -0.06345466
#> [26,] 0.1440118 -0.06044139
#> [27,] 0.1473911 -0.04713467
#> [28,] 0.1246473 -0.05412529
#> [29,] 0.1153335 -0.07589027
#> [30,] 0.1355476 -0.03683457
#> [31,] 0.1614021 -0.04176164
#> [32,] 0.1333410 -0.04606856
#> [33,] 0.1540906 -0.04781144
#> [34,] 0.1494646 -0.04525491
#> [35,] 0.1338232 -0.06644229
#> [36,] 0.1297536 -0.06902136
#> [37,] 0.1451209 -0.04859606
#> [38,] 0.1565349 -0.04034551
#> [39,] 0.1328010 -0.05803862
#> [40,] 0.1481568 -0.04481570
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5468474  0.5468474
#>  [2,]  0.4537487 -0.4537487
#>  [3,] -0.5489415  0.5489415
#>  [4,]  0.4516917 -0.4516917
#>  [5,]  0.4496384 -0.4496384
#>  [6,] -0.5431227  0.5431227
#>  [7,] -0.5508299  0.5508299
#>  [8,] -0.5516150  0.5516150
#>  [9,] -0.5509004  0.5509004
#> [10,] -0.5469235  0.5469235
#> [11,] -0.5454944  0.5454944
#> [12,] -0.5475442  0.5475442
#> [13,]  0.4535308 -0.4535308
#> [14,]  0.4528568 -0.4528568
#> [15,] -0.5485373  0.5485373
#> [16,]  0.4529094 -0.4529094
#> [17,]  0.4500906 -0.4500906
#> [18,]  0.4546833 -0.4546833
#> [19,]  0.4491817 -0.4491817
#> [20,] -0.5509549  0.5509549
#> [21,] -0.5448076  0.5448076
#> [22,] -0.5467933  0.5467933
#> [23,]  0.4550420 -0.4550420
#> [24,] -0.5476575  0.5476575
#> [25,]  0.4513542 -0.4513542
#> [26,]  0.4490640 -0.4490640
#> [27,] -0.5484787  0.5484787
#> [28,]  0.4554255 -0.4554255
#> [29,]  0.4523392 -0.4523392
#> [30,]  0.4570109 -0.4570109
#> [31,] -0.5506169  0.5506169
#> [32,]  0.4552675 -0.4552675
#> [33,] -0.5503047  0.5503047
#> [34,]  0.4514733 -0.4514733
#> [35,] -0.5498997  0.5498997
#> [36,] -0.5495308  0.5495308
#> [37,]  0.4517216 -0.4517216
#> [38,] -0.5490617  0.5490617
#> [39,] -0.5475656  0.5475656
#> [40,] -0.5480940  0.5480940
#> 
#> $L
#> [1] 0.7140283 0.7683411 0.7123123 0.7180267 0.7953163 0.7083911
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319