Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12          V13
#>  [1,] -0.4086761 -0.15443542 -0.119157906 -0.082848492 -0.022275393
#>  [2,] -0.2252151 -0.09379717 -0.154566636 -0.195254918 -0.099212296
#>  [3,] -0.2636345 -0.11724659 -0.003031208 -0.002020592 -0.002265759
#>  [4,] -0.3318659 -0.09460087 -0.111070064 -0.133026859 -0.084189044
#>  [5,] -0.3326390 -0.13772292 -0.074839551  0.002227558 -0.115495857
#>  [6,] -0.2719775 -0.13036594 -0.057944827 -0.112989906 -0.074594289
#>  [7,] -0.2081172 -0.09601601 -0.172840510 -0.046662171 -0.100264433
#>  [8,] -0.3541277 -0.16862439 -0.082427228 -0.165916010 -0.005904182
#>  [9,] -0.2714370 -0.10687690 -0.006586176 -0.146765935 -0.004204478
#> [10,] -0.3347183 -0.08781513 -0.096317592  0.021338911  0.053696397
#>                V14          V15         V16          V17          V18
#>  [1,] -0.041553990 -0.132550227 -0.03737276 -0.082779461 -0.045431263
#>  [2,] -0.127295375 -0.135292417 -0.02113439 -0.082611714  0.035738947
#>  [3,] -0.019099296 -0.103694550  0.07213452 -0.061332053 -0.047611385
#>  [4,]  0.004695950 -0.052329437 -0.02000489 -0.009279382 -0.053222249
#>  [5,]  0.041918270  0.003785325 -0.09083450 -0.075066495  0.052455095
#>  [6,] -0.052155207 -0.043315143 -0.07824185 -0.007612248  0.008964855
#>  [7,] -0.088099260  0.061202129 -0.08765858  0.111100236  0.043550812
#>  [8,] -0.071490854 -0.138926124 -0.07054192  0.100692497 -0.051737332
#>  [9,] -0.107200472  0.048966432  0.03233474 -0.054072688  0.111585213
#> [10,]  0.006556928  0.131835209  0.18480279  0.070267300  0.102303059
#>               V19         V2          V20          V21          V22         V23
#>  [1,] -0.07676551 -0.2383799 -0.006869298  0.098060180  0.069121303  0.14907065
#>  [2,]  0.01407103 -0.2731979 -0.009280393  0.065115600 -0.023576851 -0.02934216
#>  [3,] -0.01426590 -0.3051745  0.071931914  0.036025908  0.133250265  0.04157400
#>  [4,]  0.02734016 -0.3067487  0.040355174 -0.079526368 -0.007426302  0.01020713
#>  [5,]  0.01028144 -0.2646952  0.066814229  0.097463299  0.090753794  0.16487751
#>  [6,]  0.09334831 -0.2787636  0.032289221  0.202778491  0.185713269  0.17735614
#>  [7,]  0.00244177 -0.2645631  0.009961113  0.065561910  0.183455409  0.11969085
#>  [8,]  0.14126007 -0.2322558  0.047817596  0.050588110  0.150450320  0.06956141
#>  [9,]  0.09576062 -0.3757707  0.004124895 -0.003579077  0.068375686  0.16593048
#> [10,]  0.04908015 -0.3060784  0.124207952  0.235080649  0.156952078  0.26034619
#>               V24         V25          V26         V27          V28         V29
#>  [1,]  0.05211804  0.18977536  0.126365777  0.28167975  0.131444513  0.04147986
#>  [2,] -0.07641311 -0.02818626 -0.047226584 -0.00932991  0.067382207 -0.06185416
#>  [3,]  0.09947211  0.23115513  0.190452642  0.04751549  0.259161952  0.09456003
#>  [4,]  0.01965942  0.05810152  0.003838009  0.15122537 -0.001948727 -0.06864385
#>  [5,]  0.04618419  0.13761429  0.152289172  0.13273748  0.111370261  0.11731252
#>  [6,]  0.26954322  0.13811859  0.156313828  0.24015603  0.271449217  0.17101439
#>  [7,]  0.11971763  0.19200537  0.199073646  0.19947075  0.117620060  0.14864372
#>  [8,]  0.22421749  0.24231144  0.244917153  0.12970773  0.159091240  0.14056342
#>  [9,]  0.21000586  0.14405694  0.286977147  0.17597611  0.316369247  0.26175807
#> [10,]  0.30458674  0.14950479  0.164240141  0.27310930  0.168144470  0.16896190
#>               V3           V30         V31          V32          V33
#>  [1,] -0.2842540  0.1366446519  0.01707149 -0.007278319  0.062392616
#>  [2,] -0.2145130 -0.1394373382 -0.06122031 -0.125516780 -0.129239111
#>  [3,] -0.2235857  0.0006599448  0.11978731  0.035204788 -0.121398793
#>  [4,] -0.2643453  0.0672429399 -0.01682637 -0.065514732  0.013463375
#>  [5,] -0.2347841  0.1130373434  0.15538559  0.133378093  0.064891300
#>  [6,] -0.3692926  0.0739959251  0.03149089  0.040350712 -0.047692769
#>  [7,] -0.2381158  0.0515658812  0.03081493  0.058281958  0.008003001
#>  [8,] -0.2196827  0.2011348078  0.05793833 -0.019873776  0.042835745
#>  [9,] -0.2596930  0.1681096006  0.04220996  0.047761045  0.033767414
#> [10,] -0.3926521  0.1385924579  0.09090066  0.007658857 -0.031268102
#>                V34         V35          V36         V37          V38
#>  [1,]  0.109810886 -0.00808634  0.032989565  0.06908461  0.043601789
#>  [2,] -0.165134938 -0.13940257 -0.154864465 -0.07708226 -0.146733295
#>  [3,] -0.108719386  0.01224697 -0.087310626  0.05031949 -0.132471563
#>  [4,] -0.118375532 -0.04384282 -0.151447692 -0.20756159 -0.047371444
#>  [5,]  0.066865419 -0.09490576 -0.054428991 -0.04108499 -0.002862429
#>  [6,] -0.112821396 -0.13164477 -0.140806062 -0.18293098 -0.156303605
#>  [7,]  0.069279825  0.02483723 -0.005232615 -0.04597151 -0.156676963
#>  [8,] -0.052633145  0.00608919  0.044367543 -0.03719069 -0.083984794
#>  [9,] -0.004492345  0.01628987  0.081122342 -0.07840035  0.021237622
#> [10,]  0.076948618 -0.03190886  0.071456562  0.08902323 -0.063652361
#>                V39         V4         V40          V41          V42
#>  [1,] -0.169047951 -0.3296764  0.01968827  0.025357024 -0.014388649
#>  [2,] -0.073932231 -0.1736397 -0.12359391 -0.102195984 -0.117397819
#>  [3,] -0.043730772 -0.1981382 -0.17327237 -0.070597734 -0.180490881
#>  [4,] -0.089809310 -0.2634692 -0.02196741 -0.029026188 -0.173934303
#>  [5,]  0.050142958 -0.2541977 -0.06664223 -0.006164250  0.026429346
#>  [6,] -0.070807750 -0.3196646 -0.13468546 -0.061859468 -0.215424468
#>  [7,] -0.147351869 -0.2706043 -0.05438884 -0.144983133 -0.056214727
#>  [8,] -0.152193956 -0.1845261 -0.15318060  0.008463547  0.003662730
#>  [9,]  0.001730528 -0.3038515  0.02160825 -0.078396674  0.003949626
#> [10,] -0.091018867 -0.3451783 -0.09432437 -0.052504868 -0.062712906
#>                V43         V44         V45         V46        V47        V48
#>  [1,] -0.105406779 -0.21522848 -0.12634529 -0.13295703 -0.2107568 -0.2314740
#>  [2,] -0.217576126 -0.22153635 -0.23671617 -0.19663359 -0.1767845 -0.2259563
#>  [3,] -0.180958448 -0.06646432 -0.00135865 -0.17143054 -0.1282136 -0.1966184
#>  [4,] -0.065719067 -0.16320931 -0.06953509 -0.25353996 -0.1506949 -0.1568546
#>  [5,] -0.037588268 -0.08146412 -0.10847575 -0.21476506 -0.2283749 -0.2190502
#>  [6,] -0.048226417 -0.09381475 -0.11084186 -0.17547208 -0.1496710 -0.1909800
#>  [7,] -0.050888298 -0.14724416 -0.13988029 -0.15906683 -0.1422347 -0.2643013
#>  [8,] -0.027006277 -0.10707028 -0.16719277 -0.09349941 -0.2134275 -0.1140847
#>  [9,] -0.002015804 -0.16521151 -0.10520377 -0.07260705 -0.2500292 -0.2318410
#> [10,] -0.125551493  0.04022141 -0.03645047 -0.12543068 -0.1354315 -0.2554402
#>              V49         V5        V50        V51        V52        V53
#>  [1,] -0.2616246 -0.3192828 -0.3068089 -0.3892284 -0.3325720 -0.4249194
#>  [2,] -0.1746763 -0.2208993 -0.2367059 -0.3534682 -0.2091424 -0.3586028
#>  [3,] -0.2203273 -0.3037391 -0.3568954 -0.3410598 -0.3831061 -0.3596349
#>  [4,] -0.3006934 -0.1850114 -0.1758337 -0.2212594 -0.2170633 -0.3223094
#>  [5,] -0.2544951 -0.1963908 -0.2945242 -0.2920160 -0.2503945 -0.2488867
#>  [6,] -0.2494228 -0.2847488 -0.3315655 -0.4112311 -0.3048020 -0.4482666
#>  [7,] -0.2822949 -0.1293711 -0.2950697 -0.2584717 -0.2728240 -0.3883586
#>  [8,] -0.2617903 -0.1625126 -0.2783688 -0.2213895 -0.3324727 -0.2193027
#>  [9,] -0.3148239 -0.3283985 -0.4560119 -0.4018315 -0.3237326 -0.4075310
#> [10,] -0.2045180 -0.2732407 -0.3941862 -0.2564307 -0.4445900 -0.3749876
#>              V54        V55        V56        V57        V58        V59
#>  [1,] -0.3798576 -0.4204251 -0.3453001 -0.2991919 -0.4156205 -0.3191183
#>  [2,] -0.2466093 -0.2597399 -0.2617144 -0.2526448 -0.2619987 -0.2082168
#>  [3,] -0.3121233 -0.3209074 -0.3489152 -0.3235902 -0.3981588 -0.3866853
#>  [4,] -0.2429942 -0.2122436 -0.1437066 -0.3145365 -0.1667813 -0.2634609
#>  [5,] -0.3945909 -0.3223020 -0.2575953 -0.2778127 -0.2197272 -0.3916561
#>  [6,] -0.4196412 -0.3928178 -0.4090204 -0.3547427 -0.3406544 -0.3735216
#>  [7,] -0.2591164 -0.3224628 -0.3716387 -0.3132561 -0.3353029 -0.4119098
#>  [8,] -0.2728190 -0.3531792 -0.2588529 -0.3701095 -0.3382409 -0.3824675
#>  [9,] -0.2896301 -0.3225189 -0.4115571 -0.3669399 -0.4473014 -0.3453049
#> [10,] -0.4057462 -0.3851976 -0.4154230 -0.3769652 -0.4072351 -0.2798028
#>               V6        V60          V7          V8          V9
#>  [1,] -0.2360748 -0.3611176 -0.20773350 -0.20563670 -0.17588707
#>  [2,] -0.1297174 -0.2473818 -0.09699896 -0.10646983 -0.15787014
#>  [3,] -0.1739006 -0.3964644 -0.10611821 -0.14715552 -0.12680318
#>  [4,] -0.2456058 -0.3287982 -0.20109254 -0.09051411 -0.08660570
#>  [5,] -0.1312547 -0.2173911 -0.19817422 -0.13980193 -0.09032624
#>  [6,] -0.3009824 -0.4600730 -0.28381706 -0.24899226 -0.14837640
#>  [7,] -0.1536552 -0.3764306 -0.08020061 -0.22749313 -0.04502226
#>  [8,] -0.1210507 -0.3021304 -0.20269922 -0.19528937 -0.17471999
#>  [9,] -0.2618290 -0.4153749 -0.24559385 -0.20051912 -0.15682156
#> [10,] -0.2351825 -0.3932520 -0.10711664 -0.12942759 -0.06499085
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]       [,5]        [,6]
#> [1,] -0.03133034 -0.03495668 -0.03430428 -0.01775499 0.06893092  0.04417465
#> [2,]  0.09762005  0.01931865  0.02527583 -0.04875944 0.04011933 -0.12013943
#>            [,7]         [,8]       [,9]        [,10]
#> [1,] 0.05567001  0.006388911  0.1045117  0.009939176
#> [2,] 0.03194225 -0.114800781 -0.1025310 -0.058990709
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  4.913848e-04  2.810941e-03  3.314748e-03  0.0031959051  2.920078e-03
#>  [2,]  9.007008e-05  5.293971e-04  6.585077e-04  0.0006846312  6.256479e-04
#>  [3,]  3.729865e-04  2.211612e-03  2.600475e-03  0.0025365161  2.316419e-03
#>  [4,]  1.972193e-05  5.283634e-05  7.524356e-05  0.0001754684  2.907679e-04
#>  [5,]  1.989030e-04  1.144678e-03  1.352959e-03  0.0014113909  1.390304e-03
#>  [6,] -7.348760e-05 -3.803399e-04 -4.301041e-04 -0.0001933735  3.917703e-05
#>  [7,]  1.869258e-04  1.077845e-03  1.280748e-03  0.0013665502  1.361625e-03
#>  [8,]  5.167508e-06  3.267713e-05  6.865603e-05  0.0002853356  4.438032e-04
#>  [9,] -1.012888e-04 -5.037762e-04 -5.463714e-04 -0.0002743287 -8.306409e-05
#> [10,]  1.158580e-04  6.700621e-04  7.719581e-04  0.0008281295  8.407938e-04
#>                V14          V15          V16          V17          V18
#>  [1,] 2.439388e-03 0.0026870566 0.0029250279 0.0023692809 0.0024184492
#>  [2,] 4.964433e-04 0.0005462330 0.0006138265 0.0005569001 0.0005957346
#>  [3,] 1.926823e-03 0.0021009896 0.0022977622 0.0019161814 0.0019573865
#>  [4,] 3.959576e-04 0.0005219302 0.0006944002 0.0007797336 0.0008135974
#>  [5,] 1.267007e-03 0.0014794778 0.0017226043 0.0016003082 0.0016495110
#>  [6,] 2.960085e-04 0.0004832951 0.0007589985 0.0010887923 0.0011824482
#>  [7,] 1.262898e-03 0.0014822147 0.0017402522 0.0016502176 0.0017022987
#>  [8,] 5.617428e-04 0.0007628721 0.0010365918 0.0012366914 0.0013136693
#>  [9,] 8.423659e-05 0.0002079807 0.0004409898 0.0007931715 0.0008918386
#> [10,] 7.836233e-04 0.0009305271 0.0010990611 0.0010348067 0.0010682461
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0040022501  7.195034e-04 5.139431e-03 5.325587e-03 5.077137e-03
#>  [2,] 0.0009289928  1.274247e-04 1.095952e-03 1.113981e-03 1.076436e-03
#>  [3,] 0.0031709891  5.485858e-04 3.985491e-03 4.125725e-03 3.956514e-03
#>  [4,] 0.0007994681  2.533527e-05 8.104527e-04 7.875464e-04 7.343666e-04
#>  [5,] 0.0022165922  2.910251e-04 2.595485e-03 2.633815e-03 2.516455e-03
#>  [6,] 0.0007927791 -1.186344e-04 4.819702e-04 4.165991e-04 4.545015e-04
#>  [7,] 0.0022170044  2.756696e-04 2.559991e-03 2.588228e-03 2.473043e-03
#>  [8,] 0.0012167788  5.131104e-06 1.075953e-03 1.010078e-03 9.933461e-04
#>  [9,] 0.0004963204 -1.495722e-04 9.955419e-05 1.110718e-05 9.502313e-05
#> [10,] 0.0013687105  1.767545e-04 1.563438e-03 1.563874e-03 1.472525e-03
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0049958690 0.0049618326 0.0047918287 0.0053791016 0.0061779893
#>  [2,] 0.0010795353 0.0010739501 0.0010196595 0.0011159172 0.0013149031
#>  [3,] 0.0039400936 0.0039446557 0.0038316170 0.0043083874 0.0049916830
#>  [4,] 0.0008281636 0.0009506462 0.0010461912 0.0011052281 0.0010923948
#>  [5,] 0.0025899753 0.0026806861 0.0026749106 0.0029237099 0.0032402725
#>  [6,] 0.0006853111 0.0009173129 0.0010684629 0.0009787283 0.0008287593
#>  [7,] 0.0025573125 0.0026730727 0.0026967671 0.0029476119 0.0032482370
#>  [8,] 0.0011717196 0.0013450799 0.0014501299 0.0014782637 0.0014900477
#>  [9,] 0.0003016511 0.0004702383 0.0006066983 0.0005052157 0.0003852347
#> [10,] 0.0015449056 0.0016359580 0.0016464173 0.0018227520 0.0020131273
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0069062998 0.0059014937  6.796372e-04 0.0053566798 0.0044564185
#>  [2,] 0.0014197098 0.0012142019  1.245433e-04 0.0010659422 0.0009157752
#>  [3,] 0.0055681890 0.0047604298  5.260865e-04 0.0042921068 0.0035765450
#>  [4,] 0.0010565934 0.0010752958  3.420189e-05 0.0010399961 0.0010013337
#>  [5,] 0.0035229750 0.0031471581  2.884726e-04 0.0028954351 0.0025410541
#>  [6,] 0.0006931524 0.0009336641 -7.992109e-05 0.0008816588 0.0009928442
#>  [7,] 0.0035140117 0.0031752882  2.744586e-04 0.0029310320 0.0025868680
#>  [8,] 0.0014680983 0.0015221021  2.920570e-05 0.0014277914 0.0014402308
#>  [9,] 0.0001815691 0.0003958518 -1.146442e-04 0.0003616502 0.0005799017
#> [10,] 0.0022329516 0.0020706667  1.767898e-04 0.0019322731 0.0017108462
#>                V32          V33          V34          V35          V36
#>  [1,] 0.0042517012 0.0034825489 0.0026445210 0.0023190499 0.0023598086
#>  [2,] 0.0008666234 0.0007226455 0.0005233813 0.0004071674 0.0004421165
#>  [3,] 0.0033849995 0.0027403395 0.0020571509 0.0017653089 0.0018183683
#>  [4,] 0.0008733056 0.0010497544 0.0011911467 0.0012114921 0.0012535527
#>  [5,] 0.0023575033 0.0022155295 0.0020113207 0.0018996394 0.0019874380
#>  [6,] 0.0007992217 0.0013032493 0.0016927921 0.0017960213 0.0019197092
#>  [7,] 0.0023799572 0.0022740647 0.0021143322 0.0020149074 0.0020980081
#>  [8,] 0.0012483209 0.0015468765 0.0017643073 0.0017961932 0.0019125035
#>  [9,] 0.0004366739 0.0008565916 0.0011891027 0.0012511275 0.0013933335
#> [10,] 0.0015676961 0.0015494824 0.0014778541 0.0014223259 0.0014835218
#>                V37          V38          V39            V4          V40
#>  [1,] 0.0031739035 0.0036364671 0.0037760310  8.669077e-04 0.0027811083
#>  [2,] 0.0006621497 0.0007443127 0.0007870079  1.512164e-04 0.0006247139
#>  [3,] 0.0025001640 0.0028488053 0.0029455112  6.779208e-04 0.0022051476
#>  [4,] 0.0011288485 0.0007889353 0.0006045332  1.547124e-05 0.0006684293
#>  [5,] 0.0022179814 0.0020869214 0.0019695087  3.455843e-04 0.0016747935
#>  [6,] 0.0015548712 0.0007902776 0.0004094189 -1.442034e-04 0.0007919464
#>  [7,] 0.0022989608 0.0021056268 0.0019534634  3.219364e-04 0.0016948612
#>  [8,] 0.0017566112 0.0011893713 0.0008958138 -8.716131e-06 0.0010712352
#>  [9,] 0.0011343270 0.0004733452 0.0001700138 -1.918599e-04 0.0005736497
#> [10,] 0.0015714665 0.0014150127 0.0012894777  2.067404e-04 0.0011357442
#>                V41          V42           V43           V44           V45
#>  [1,] 0.0026858412 2.993581e-03  0.0030133210  0.0028951773  3.405580e-03
#>  [2,] 0.0005629629 5.782434e-04  0.0005539284  0.0005303780  5.939826e-04
#>  [3,] 0.0021527565 2.355990e-03  0.0023412001  0.0022676164  2.658040e-03
#>  [4,] 0.0005242106 3.766279e-04  0.0001751677  0.0001272298  2.923293e-05
#>  [5,] 0.0015144901 1.480116e-03  0.0012885944  0.0011975933  1.298240e-03
#>  [6,] 0.0005345792 1.651774e-04 -0.0002517072 -0.0003370736 -7.005482e-04
#>  [7,] 0.0015304715 1.470626e-03  0.0012478460  0.0011505454  1.223166e-03
#>  [8,] 0.0008539312 5.742311e-04  0.0002040494  0.0001079213 -1.073606e-04
#>  [9,] 0.0003844013 2.767517e-05 -0.0003492229 -0.0004199031 -7.594266e-04
#> [10,] 0.0010213887 9.716691e-04  0.0007884467  0.0007109751  7.592494e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  2.643046e-03  1.652996e-03  1.142019e-03  7.722273e-04  9.966258e-04
#>  [2,]  4.578480e-04  3.092943e-04  2.127456e-04  1.511017e-04  1.931398e-04
#>  [3,]  2.070048e-03  1.322098e-03  9.171522e-04  6.186207e-04  7.900024e-04
#>  [4,]  2.153660e-05  4.130893e-05  2.225216e-05  7.555826e-06  4.850272e-05
#>  [5,]  1.019190e-03  6.896216e-04  4.744497e-04  3.124290e-04  4.293432e-04
#>  [6,] -5.423208e-04 -2.327295e-04 -1.541105e-04 -1.246367e-04 -8.783909e-05
#>  [7,]  9.585865e-04  6.537939e-04  4.467966e-04  2.945240e-04  4.062876e-04
#>  [8,] -7.571777e-05  4.384668e-05  2.226847e-05  5.702006e-06  5.731524e-05
#>  [9,] -5.928884e-04 -2.659225e-04 -1.947707e-04 -1.346629e-04 -1.322120e-04
#> [10,]  6.076829e-04  4.288029e-04  2.974816e-04  1.893640e-04  2.594860e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.128620e-04  2.139502e-04  2.202294e-04  1.163439e-04  1.388991e-04
#>  [2,]  3.965852e-05  4.115320e-05  3.965231e-05  2.097577e-05  2.895013e-05
#>  [3,]  1.705040e-04  1.655900e-04  1.694207e-04  9.079579e-05  1.082619e-04
#>  [4,]  2.055175e-05  5.056941e-06  6.147519e-06  1.335882e-05  1.618223e-05
#>  [5,]  1.016810e-04  8.548238e-05  8.729028e-05  5.541515e-05  6.519732e-05
#>  [6,] -2.442261e-06 -3.032877e-05 -3.808316e-05  2.009946e-06  2.280886e-07
#>  [7,]  9.909573e-05  8.090632e-05  8.219951e-05  5.432475e-05  6.359713e-05
#>  [8,]  2.923770e-05  2.492002e-06 -1.254815e-06  1.747371e-05  1.919877e-05
#>  [9,] -1.065105e-05 -3.950281e-05 -4.774307e-05 -6.351455e-06 -8.528089e-06
#> [10,]  6.521779e-05  5.099646e-05  5.265384e-05  3.593306e-05  4.024521e-05
#>                V55           V56          V57           V58           V59
#>  [1,] 1.034967e-04  9.949603e-05 7.374656e-05  1.073710e-04  1.109851e-04
#>  [2,] 1.973032e-05  2.031255e-05 1.460266e-05  2.122300e-05  2.249510e-05
#>  [3,] 8.112441e-05  7.878762e-05 5.801928e-05  8.299024e-05  8.521023e-05
#>  [4,] 1.662434e-05  1.255360e-05 1.700096e-05  1.063485e-05  1.139084e-05
#>  [5,] 5.333058e-05  4.809352e-05 4.302644e-05  4.973307e-05  5.163690e-05
#>  [6,] 8.501557e-06  2.673091e-06 1.923101e-05 -2.102143e-06 -8.515904e-07
#>  [7,] 5.299564e-05  4.686901e-05 4.322477e-05  4.819175e-05  4.986290e-05
#>  [8,] 2.178124e-05  1.546172e-05 2.537908e-05  1.339394e-05  1.431122e-05
#>  [9,] 3.865983e-07 -4.781962e-06 1.037631e-05 -8.324195e-06 -9.035101e-06
#> [10,] 3.473603e-05  2.944513e-05 2.779513e-05  3.104421e-05  3.247902e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.134222e-03  8.236492e-05  1.299390e-03  1.693680e-03  2.673115e-03
#>  [2,]  2.390323e-04  1.466405e-05  2.626032e-04  3.000610e-04  4.801381e-04
#>  [3,]  8.887259e-04  6.193952e-05  1.024880e-03  1.313987e-03  2.086389e-03
#>  [4,]  1.313262e-04  7.346791e-06  1.461554e-04  1.124445e-04  4.048501e-05
#>  [5,]  5.512995e-04  3.652320e-05  6.368032e-04  7.565423e-04  1.054223e-03
#>  [6,]  5.042390e-05 -3.832598e-06  5.363116e-05 -9.989744e-05 -4.484642e-04
#>  [7,]  5.393868e-04  3.515927e-05  6.194014e-04  7.275899e-04  9.910806e-04
#>  [8,]  1.942153e-04  7.193561e-06  2.148402e-04  1.388804e-04 -3.434575e-05
#>  [9,] -1.319788e-05 -1.089511e-05 -3.093472e-05 -1.947899e-04 -5.661119e-04
#> [10,]  3.414249e-04  2.241185e-05  3.995863e-04  4.651785e-04  6.100274e-04
#> 
#> $vW[[2]]
#>             [,1]         [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.04504154 -0.005833989 -0.04627218 -0.01457528 -0.05498497 -0.04862483
#> [2,]  0.04504154  0.005833989  0.04627218  0.01457528  0.05498497  0.04862483
#>             [,7]        [,8]        [,9]      [,10]
#> [1,] -0.04764389 -0.05065101 -0.06152516 -0.0685887
#> [2,]  0.04764389  0.05065101  0.06152516  0.0685887
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.53445225  0.08819050  0.33101013 -0.03942657  0.53655024  0.38672278
#>  [7]  0.47432696  0.45936505  0.69424034  0.61811669
#> 
#> $B[[2]]
#> [1] 0.07802361 0.01401943
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.007927876 0.001651488 0.006298059 0.001877816 0.004611783 0.002022652
#>  [7] 0.004677378 0.002744208 0.001240361 0.003045071
#> 
#> $vB[[2]]
#> [1] -0.08259638  0.08259638
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [2,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#>  [3,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [4,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#>  [5,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#>  [6,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [7,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#>  [8,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#>  [9,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [10,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [11,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#> [12,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [13,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [14,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [15,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [16,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [17,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [18,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [19,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [20,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [21,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [22,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [23,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [24,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [25,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [26,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [27,] 0.0180 0.2918 0.3273 0.3035 0.3033 0.2587 0.1682 0.1308 0.2803 0.4519
#> [28,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [29,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [30,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [31,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [32,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [33,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [34,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [35,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [36,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [37,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [38,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [39,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [40,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [2,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#>  [3,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [4,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#>  [5,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#>  [6,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [7,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#>  [8,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#>  [9,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [10,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [11,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#> [12,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [13,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [14,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [15,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [16,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [17,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [18,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [19,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [20,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [21,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [22,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [23,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [24,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [25,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [26,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [27,] 0.6641 0.0444 0.7683 0.6960 0.4393 0.2432 0.2886 0.4974 0.8172 1.0000
#> [28,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [29,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [30,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [31,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [32,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [33,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [34,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [35,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [36,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [37,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [38,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [39,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [40,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [2,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#>  [3,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [4,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#>  [5,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#>  [6,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [7,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#>  [8,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#>  [9,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [10,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [11,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#> [12,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [13,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [14,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [15,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [16,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [17,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [18,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [19,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [20,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [21,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [22,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [23,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [24,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [25,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [26,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [27,] 0.9238 0.8519 0.0476 0.7722 0.5772 0.5190 0.6824 0.6220 0.5054 0.3578
#> [28,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [29,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [30,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [31,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [32,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [33,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [34,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [35,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [36,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [37,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [38,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [39,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [40,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [2,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#>  [3,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [4,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#>  [5,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#>  [6,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [7,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#>  [8,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#>  [9,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [10,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [11,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#> [12,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [13,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [14,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [15,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [16,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [17,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [18,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [19,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [20,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [21,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [22,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [23,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [24,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [25,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [26,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [27,] 0.3809 0.3813 0.3359 0.0698 0.2771 0.3648 0.3834 0.3453 0.2096 0.1031
#> [28,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [29,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [30,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [31,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [32,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [33,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [34,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [35,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [36,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [37,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [38,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [39,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [40,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [2,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#>  [3,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [4,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#>  [5,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#>  [6,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [7,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#>  [8,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#>  [9,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [10,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [11,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#> [12,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [13,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [14,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [15,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [16,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [17,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [18,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [19,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [20,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [21,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [22,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [23,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [24,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [25,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [26,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [27,] 0.0798 0.0701 0.0526 0.0241 0.1615 0.0117 0.0122 0.0122 0.0114 0.0098
#> [28,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [29,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [30,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [31,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [32,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [33,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [34,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [35,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [36,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [37,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [38,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [39,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [40,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [2,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#>  [3,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [4,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#>  [5,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#>  [6,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [7,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#>  [8,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#>  [9,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [10,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [11,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#> [12,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [13,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [14,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [15,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [16,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [17,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [18,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [19,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [20,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [21,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [22,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [23,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [24,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [25,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [26,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [27,] 0.0027 0.0025 0.0026 0.0050 0.0073 0.0887 0.0022 0.0596 0.1071 0.3175
#> [28,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [29,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [30,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [31,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [32,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [33,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [34,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [35,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [36,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [37,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [38,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [39,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [40,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5386164 0.1115319 0.4542144 0.1858016 0.6218369 0.3795036 0.5106194
#>  [2,] 0.6482582 0.2001068 0.5848894 0.2714891 0.6923863 0.6172788 0.6667031
#>  [3,] 0.6631406 0.2934288 0.6278564 0.3690399 0.7466333 0.7119189 0.7355657
#>  [4,] 0.7086301 0.3338266 0.6637881 0.3908529 0.7500112 0.7442314 0.7532525
#>  [5,] 0.7283827 0.1758137 0.5756662 0.2918755 0.7479562 0.6251678 0.6935948
#>  [6,] 0.6718904 0.1581321 0.5612263 0.2450640 0.6863004 0.5619230 0.6351051
#>  [7,] 0.7502553 0.3392412 0.6953817 0.4354939 0.7494160 0.7287676 0.7576644
#>  [8,] 0.7797356 0.2931025 0.7062790 0.4214537 0.7897523 0.7566247 0.7857611
#>  [9,] 0.7170178 0.2066224 0.6873563 0.3214800 0.7600486 0.7367187 0.7437297
#> [10,] 0.6923381 0.2053962 0.6249297 0.3175825 0.7115814 0.6637284 0.7204641
#> [11,] 0.6146428 0.2416220 0.6278066 0.3312360 0.7145330 0.6986822 0.7165730
#> [12,] 0.5314654 0.2549711 0.5180556 0.3076253 0.6242998 0.5363362 0.6110515
#> [13,] 0.7652131 0.2365189 0.6921056 0.3795383 0.8039734 0.7431545 0.7754618
#> [14,] 0.6931542 0.1264125 0.5701732 0.2522615 0.7274247 0.5405418 0.6577313
#> [15,] 0.7562380 0.2753073 0.6982593 0.3849126 0.7853854 0.7534486 0.7583876
#> [16,] 0.7296952 0.2523817 0.7128708 0.3806020 0.7724188 0.7631892 0.7785444
#> [17,] 0.7549206 0.2671061 0.7290610 0.3830931 0.8011685 0.8161515 0.7960015
#> [18,] 0.6363100 0.1645719 0.5271174 0.2480700 0.7181135 0.5467123 0.6447737
#> [19,] 0.7341352 0.2649696 0.6869019 0.3756755 0.7705981 0.7407109 0.7505211
#> [20,] 0.5459665 0.1259953 0.4572950 0.1956691 0.5988291 0.3611980 0.5145602
#> [21,] 0.6569829 0.1486468 0.6060310 0.2572938 0.7352920 0.6430541 0.6861052
#> [22,] 0.7946452 0.3445962 0.7575046 0.4447324 0.8181334 0.8374956 0.8151877
#> [23,] 0.7267954 0.1461627 0.6150205 0.2813011 0.7484421 0.6174904 0.7019933
#> [24,] 0.6865667 0.3166485 0.6798782 0.3658162 0.7221356 0.7142045 0.7341818
#> [25,] 0.8306463 0.2463170 0.7358298 0.3764389 0.8349484 0.7941265 0.8243576
#> [26,] 0.7902820 0.2483464 0.7353509 0.3705818 0.8069758 0.7913858 0.7920178
#> [27,] 0.7286212 0.2160802 0.6279012 0.3335092 0.7728591 0.6821397 0.7279159
#> [28,] 0.7075597 0.1979804 0.6688114 0.3235502 0.7489917 0.7336851 0.7261379
#> [29,] 0.6908480 0.2270148 0.6650868 0.3437847 0.7608447 0.7375944 0.7695709
#> [30,] 0.7415574 0.3353814 0.7266415 0.4257922 0.7835189 0.8105966 0.8010743
#> [31,] 0.6344475 0.1655646 0.5435972 0.2806962 0.6997586 0.5798515 0.6429332
#> [32,] 0.7365827 0.2499549 0.7011067 0.3616655 0.7910739 0.7858793 0.7753077
#> [33,] 0.7516877 0.2754713 0.7210497 0.3900783 0.7860213 0.7896273 0.7822738
#> [34,] 0.6824116 0.3052364 0.6516357 0.3828184 0.7389714 0.7311351 0.7472860
#> [35,] 0.7611667 0.2201796 0.6415361 0.3484107 0.7676283 0.6755622 0.7559115
#> [36,] 0.7332663 0.2390974 0.6593566 0.3617585 0.7611289 0.7158784 0.7263258
#> [37,] 0.6695424 0.2162988 0.6456810 0.3321556 0.7492470 0.7091187 0.7498623
#> [38,] 0.7632376 0.2044243 0.6595597 0.3142613 0.7806313 0.6983420 0.7624857
#> [39,] 0.6922764 0.2472467 0.6331726 0.3288981 0.7562858 0.7059513 0.7356789
#> [40,] 0.6872903 0.1940499 0.6623356 0.3116034 0.7587370 0.7295894 0.7127470
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5295506 0.6901353 0.7643353
#>  [2,] 0.7123857 0.7813726 0.8239242
#>  [3,] 0.7441523 0.8083173 0.8857543
#>  [4,] 0.7637619 0.8108353 0.8901763
#>  [5,] 0.7268313 0.8155622 0.8529564
#>  [6,] 0.6684135 0.7841534 0.8324933
#>  [7,] 0.7742362 0.8434008 0.8464111
#>  [8,] 0.8030087 0.8777936 0.8738101
#>  [9,] 0.7612048 0.8371962 0.9077195
#> [10,] 0.7432516 0.8356390 0.8923904
#> [11,] 0.7118963 0.8020099 0.9007075
#> [12,] 0.5939397 0.6864920 0.7994706
#> [13,] 0.7957615 0.8798373 0.8899181
#> [14,] 0.6938103 0.8263907 0.8309474
#> [15,] 0.7823508 0.8658902 0.8728182
#> [16,] 0.7893779 0.8618254 0.9086234
#> [17,] 0.8212835 0.8854545 0.9307858
#> [18,] 0.6537380 0.7595447 0.8425131
#> [19,] 0.7730412 0.8548204 0.8725503
#> [20,] 0.5054633 0.6911406 0.7767522
#> [21,] 0.7329385 0.8134528 0.8679064
#> [22,] 0.8417546 0.8872950 0.9100498
#> [23,] 0.7383304 0.8475573 0.8503987
#> [24,] 0.7359025 0.8142043 0.8611878
#> [25,] 0.8427783 0.9074269 0.9237129
#> [26,] 0.8137257 0.8888685 0.9182910
#> [27,] 0.7548123 0.8445713 0.8645533
#> [28,] 0.7365249 0.8369782 0.8951385
#> [29,] 0.7767153 0.8541192 0.9159016
#> [30,] 0.8045472 0.8590234 0.9247708
#> [31,] 0.6807256 0.7763033 0.8087825
#> [32,] 0.8030505 0.8704826 0.9173873
#> [33,] 0.7951687 0.8686663 0.8901063
#> [34,] 0.7677696 0.8299351 0.8854936
#> [35,] 0.7724667 0.8602080 0.8660992
#> [36,] 0.7591193 0.8358609 0.8709224
#> [37,] 0.7496889 0.8325704 0.9101915
#> [38,] 0.7919832 0.8473844 0.8892495
#> [39,] 0.7539726 0.8309681 0.8981131
#> [40,] 0.7297147 0.8351495 0.8856108
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4156896 0.5843104
#>  [2,] 0.4052141 0.5947859
#>  [3,] 0.4003845 0.5996155
#>  [4,] 0.3962800 0.6037200
#>  [5,] 0.4005868 0.5994132
#>  [6,] 0.4049634 0.5950366
#>  [7,] 0.3939923 0.6060077
#>  [8,] 0.3923404 0.6076596
#>  [9,] 0.3970377 0.6029623
#> [10,] 0.4009314 0.5990686
#> [11,] 0.4039576 0.5960424
#> [12,] 0.4118884 0.5881116
#> [13,] 0.3939147 0.6060853
#> [14,] 0.4034833 0.5965167
#> [15,] 0.3943042 0.6056958
#> [16,] 0.3949523 0.6050477
#> [17,] 0.3930712 0.6069288
#> [18,] 0.4062234 0.5937766
#> [19,] 0.3959882 0.6040118
#> [20,] 0.4147430 0.5852570
#> [21,] 0.4040577 0.5959423
#> [22,] 0.3889362 0.6110638
#> [23,] 0.4001624 0.5998376
#> [24,] 0.3979060 0.6020940
#> [25,] 0.3883697 0.6116303
#> [26,] 0.3910818 0.6089182
#> [27,] 0.3984202 0.6015798
#> [28,] 0.3988990 0.6011010
#> [29,] 0.3987465 0.6012535
#> [30,] 0.3924574 0.6075426
#> [31,] 0.4076681 0.5923319
#> [32,] 0.3952023 0.6047977
#> [33,] 0.3937078 0.6062922
#> [34,] 0.3991328 0.6008672
#> [35,] 0.3960482 0.6039518
#> [36,] 0.3972176 0.6027824
#> [37,] 0.4003750 0.5996250
#> [38,] 0.3949755 0.6050245
#> [39,] 0.3993870 0.6006130
#> [40,] 0.4003183 0.5996817
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]       [,2]        [,3]       [,4]      [,5]       [,6]
#>  [1,] 0.1547737 -2.0751883 -0.18365695 -1.4775246 0.4973521 -0.4916558
#>  [2,] 0.6113918 -1.3856273  0.34287796 -0.9870806 0.8112990  0.4780140
#>  [3,] 0.6773215 -0.8787888  0.52303185 -0.5363380 1.0807365  0.9047221
#>  [4,] 0.8887400 -0.6909281  0.68022155 -0.4437285 1.0986718  1.0680788
#>  [5,] 0.9864325 -1.5449718  0.30500737 -0.8862926 1.0877418  0.5115415
#>  [6,] 0.7167476 -1.6721927  0.24614051 -1.1251135 0.7828796  0.2489701
#>  [7,] 1.0999741 -0.6666774  0.82540130 -0.2594703 1.0955001  0.9883787
#>  [8,] 1.2641262 -0.8803635  0.87738014 -0.3168084 1.3234327  1.1342627
#>  [9,] 0.9297168 -1.3454062  0.78778850 -0.7469785 1.1529462  1.0289831
#> [10,] 0.8110730 -1.3529027  0.51052550 -0.7649038 0.9030767  0.6799537
#> [11,] 0.4668708 -1.1438076  0.52281836 -0.7025999 0.9175027  0.8410305
#> [12,] 0.1260281 -1.0722729  0.07225393 -0.8112447 0.5078393  0.1456016
#> [13,] 1.1814763 -1.1718601  0.80998180 -0.4915082 1.4113157  1.0624293
#> [14,] 0.8149074 -1.9330579  0.28255784 -1.0865873 0.9815957  0.1625239
#> [15,] 1.1321637 -0.9678600  0.83902233 -0.4687482 1.2973309  1.1170901
#> [16,] 0.9930768 -1.0859499  0.90936769 -0.4869937 1.2220196  1.1702447
#> [17,] 1.1250302 -1.0093551  0.98986359 -0.4764401 1.3936135  1.4904878
#> [18,] 0.5593841 -1.6245968  0.10857603 -1.1089321 0.9351230  0.1873957
#> [19,] 1.0157055 -1.0202969  0.78567472 -0.5079446 1.2116916  1.0496668
#> [20,] 0.1843865 -1.9368414 -0.17123734 -1.4135860 0.4005887 -0.5701683
#> [21,] 0.6498776 -1.7452538  0.43065891 -1.0600819 1.0216403  0.5886445
#> [22,] 1.3531566 -0.6428784  1.13904662 -0.2219774 1.5037518  1.6397112
#> [23,] 0.9784241 -1.7650200  0.46846545 -0.9380169 1.0903205  0.4789098
#> [24,] 0.7841167 -0.7692170  0.75321210 -0.5502079 0.9550799  0.9158927
#> [25,] 1.5902150 -1.1183527  1.02440532 -0.5046909 1.6211114  1.3499810
#> [26,] 1.3266260 -1.1074509  1.02194294 -0.5297217 1.4304779  1.3332987
#> [27,] 0.9876385 -1.2886569  0.52322343 -0.6923559 1.2245264  0.7636227
#> [28,] 0.8835612 -1.3989652  0.70281407 -0.7375038 1.0932421  1.0134005
#> [29,] 0.8040867 -1.2252449  0.68604601 -0.6464732 1.1573162  1.0335025
#> [30,] 1.0540789 -0.6839450  0.97764892 -0.2990399 1.2862921  1.4538912
#> [31,] 0.5513444 -1.6173936  0.17483269 -0.9410109 0.8461487  0.3221639
#> [32,] 1.0282820 -1.0988530  0.85257366 -0.5681426 1.3314107  1.3002631
#> [33,] 1.1076337 -0.9670380  0.94967463 -0.4469832 1.3011072  1.3226803
#> [34,] 0.7648768 -0.8224852  0.62623673 -0.4776026 1.0406292  1.0003894
#> [35,] 1.1590867 -1.2646200  0.58203764 -0.6260327 1.1949674  0.7334512
#> [36,] 1.0112581 -1.1576343  0.66042838 -0.5677397 1.1588788  0.9241080
#> [37,] 0.7061161 -1.2873671  0.60010777 -0.6984519 1.0946004  0.8911079
#> [38,] 1.1705120 -1.3588682  0.66133293 -0.7802721 1.2693489  0.8394148
#> [39,] 0.8107833 -1.1133509  0.54585182 -0.7131730 1.1324229  0.8758010
#> [40,] 0.7874813 -1.4239067  0.67371981 -0.7926339 1.1457674  0.9925403
#>             [,7]       [,8]      [,9]    [,10]
#>  [1,] 0.04248390 0.11834035 0.8007517 1.176597
#>  [2,] 0.69331101 0.90699923 1.2736831 1.543164
#>  [3,] 1.02304687 1.06766316 1.4391131 2.048088
#>  [4,] 1.11603475 1.17341605 1.4554469 2.092544
#>  [5,] 0.81697971 0.97860466 1.4865653 1.757980
#>  [6,] 0.55418092 0.70101808 1.2900368 1.603402
#>  [7,] 1.13991675 1.23238747 1.6837523 1.706726
#>  [8,] 1.29956132 1.40520629 1.9717002 1.935075
#>  [9,] 1.06544483 1.15929640 1.6375125 2.286103
#> [10,] 0.94676485 1.06293809 1.6261317 2.115394
#> [11,] 0.92752547 0.90461165 1.3989037 2.205110
#> [12,] 0.45173447 0.38027621 0.7837695 1.382989
#> [13,] 1.23941277 1.36001087 1.9908906 2.089905
#> [14,] 0.65320045 0.81799387 1.5602602 1.592357
#> [15,] 1.14385995 1.27941854 1.8650990 1.926110
#> [16,] 1.25720390 1.32117972 1.8305344 2.296941
#> [17,] 1.36148835 1.52506787 2.0451286 2.598823
#> [18,] 0.59614454 0.63551103 1.1501850 1.677047
#> [19,] 1.10139354 1.22556361 1.7729197 1.923698
#> [20,] 0.05825739 0.02185417 0.8054571 1.246839
#> [21,] 0.78197318 1.00958294 1.4726034 1.882573
#> [22,] 1.48407752 1.67134145 2.0634040 2.314243
#> [23,] 0.85680777 1.03730899 1.7155699 1.737731
#> [24,] 1.01594408 1.02477924 1.4775637 1.825191
#> [25,] 1.54615399 1.67904734 2.2826146 2.493898
#> [26,] 1.33713119 1.47440286 2.0792350 2.419351
#> [27,] 0.98407420 1.12444520 1.6926420 1.853634
#> [28,] 0.97511535 1.02798432 1.6359144 2.144338
#> [29,] 1.20588979 1.24662603 1.7672812 2.387922
#> [30,] 1.39302243 1.41496073 1.8072020 2.509006
#> [31,] 0.58811783 0.75710822 1.2442522 1.442119
#> [32,] 1.23852800 1.40547029 1.9052326 2.407367
#> [33,] 1.27896622 1.35636765 1.8892178 2.091827
#> [34,] 1.08418950 1.19575959 1.5851674 2.045515
#> [35,] 1.13039322 1.22229255 1.8170189 1.866901
#> [36,] 0.97606037 1.14785698 1.6277482 1.909139
#> [37,] 1.09787817 1.09695402 1.6039550 2.315975
#> [38,] 1.16635602 1.33692107 1.7142322 2.083098
#> [39,] 1.02362918 1.11991310 1.5925043 2.176433
#> [40,] 0.90876318 0.99317568 1.6225717 2.046671
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.1203823 0.2201112
#>  [2,] -0.1521555 0.2316302
#>  [3,] -0.1667008 0.2371624
#>  [4,] -0.1763971 0.2445926
#>  [5,] -0.1571165 0.2459041
#>  [6,] -0.1480160 0.2368103
#>  [7,] -0.1801869 0.2503746
#>  [8,] -0.1817184 0.2557664
#>  [9,] -0.1711907 0.2466328
#> [10,] -0.1630108 0.2385751
#> [11,] -0.1599191 0.2290828
#> [12,] -0.1396946 0.2164697
#> [13,] -0.1761646 0.2547220
#> [14,] -0.1476673 0.2433050
#> [15,] -0.1771760 0.2520790
#> [16,] -0.1778999 0.2486425
#> [17,] -0.1832030 0.2512178
#> [18,] -0.1433172 0.2362827
#> [19,] -0.1741133 0.2480960
#> [20,] -0.1219275 0.2224644
#> [21,] -0.1534382 0.2351480
#> [22,] -0.1929976 0.2587883
#> [23,] -0.1574902 0.2472985
#> [24,] -0.1718285 0.2423693
#> [25,] -0.1872640 0.2669065
#> [26,] -0.1838674 0.2588996
#> [27,] -0.1646346 0.2474173
#> [28,] -0.1671495 0.2429053
#> [29,] -0.1691749 0.2415158
#> [30,] -0.1863763 0.2506181
#> [31,] -0.1445122 0.2291015
#> [32,] -0.1773806 0.2481156
#> [33,] -0.1802610 0.2514920
#> [34,] -0.1712148 0.2378648
#> [35,] -0.1688867 0.2530717
#> [36,] -0.1706412 0.2464309
#> [37,] -0.1647594 0.2391433
#> [38,] -0.1713905 0.2550546
#> [39,] -0.1661965 0.2418235
#> [40,] -0.1637048 0.2404342
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5843104 -0.5843104
#>  [2,]  0.5947859 -0.5947859
#>  [3,]  0.5996155 -0.5996155
#>  [4,] -0.3962800  0.3962800
#>  [5,] -0.4005868  0.4005868
#>  [6,]  0.5950366 -0.5950366
#>  [7,] -0.3939923  0.3939923
#>  [8,] -0.3923404  0.3923404
#>  [9,]  0.6029623 -0.6029623
#> [10,] -0.4009314  0.4009314
#> [11,]  0.5960424 -0.5960424
#> [12,]  0.5881116 -0.5881116
#> [13,] -0.3939147  0.3939147
#> [14,]  0.5965167 -0.5965167
#> [15,]  0.6056958 -0.6056958
#> [16,] -0.3949523  0.3949523
#> [17,]  0.6069288 -0.6069288
#> [18,] -0.4062234  0.4062234
#> [19,]  0.6040118 -0.6040118
#> [20,]  0.5852570 -0.5852570
#> [21,]  0.5959423 -0.5959423
#> [22,] -0.3889362  0.3889362
#> [23,]  0.5998376 -0.5998376
#> [24,] -0.3979060  0.3979060
#> [25,] -0.3883697  0.3883697
#> [26,]  0.6089182 -0.6089182
#> [27,]  0.6015798 -0.6015798
#> [28,]  0.6011010 -0.6011010
#> [29,] -0.3987465  0.3987465
#> [30,] -0.3924574  0.3924574
#> [31,]  0.5923319 -0.5923319
#> [32,]  0.6047977 -0.6047977
#> [33,]  0.6062922 -0.6062922
#> [34,] -0.3991328  0.3991328
#> [35,] -0.3960482  0.3960482
#> [36,]  0.6027824 -0.6027824
#> [37,] -0.4003750  0.4003750
#> [38,] -0.3949755  0.3949755
#> [39,]  0.6006130 -0.6006130
#> [40,]  0.5996817 -0.5996817
#> 
#> $L
#> [1] 0.6940115 0.6910821 0.7113118 0.7633986 0.8043216 0.7397288
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4202899