Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13          V14
#>  [1,] -0.3272122 -0.09967669 -0.14518219 -0.05445670 -0.02178101 -0.179294014
#>  [2,] -0.3420481 -0.12304206 -0.05376714  0.01678124 -0.11911762  0.038772868
#>  [3,] -0.2656279 -0.03800116 -0.13711114 -0.04983622 -0.11269683 -0.130970205
#>  [4,] -0.2608805  0.07649813 -0.04033149 -0.06932402  0.02226316  0.003238360
#>  [5,] -0.2472259 -0.21561113 -0.09899098 -0.15926415 -0.23520221 -0.151338639
#>  [6,] -0.4610262 -0.20281285 -0.13088699 -0.08642050 -0.01678214 -0.007910247
#>  [7,] -0.1576939 -0.16009071 -0.12302439 -0.05461432 -0.08330288 -0.171957907
#>  [8,] -0.3944322 -0.13688903 -0.05537516 -0.15986311 -0.03408445 -0.012676265
#>  [9,] -0.2794367 -0.07791439 -0.14901439 -0.14941462 -0.12471478 -0.127405124
#> [10,] -0.3533397 -0.21003393 -0.10550288 -0.19116680 -0.03182126 -0.102788106
#>                V15          V16         V17         V18           V19
#>  [1,] -0.046731230 -0.046503147 -0.08605212  0.01422490  0.0007237469
#>  [2,] -0.100245601 -0.005817161 -0.07611832  0.03645368  0.0270190771
#>  [3,] -0.081351803 -0.035104575 -0.02037148 -0.07939386  0.0264630846
#>  [4,]  0.038465795 -0.048195944  0.14566450  0.08354768  0.2112230647
#>  [5,] -0.158760911 -0.036668277 -0.15808898 -0.17541690 -0.1032497684
#>  [6,] -0.100734150 -0.100673989 -0.05612471  0.07548603  0.0850158118
#>  [7,] -0.198308078 -0.159886028 -0.15728160 -0.06975134 -0.0751834351
#>  [8,] -0.006033994 -0.090729933  0.06805276  0.07889020  0.0400126941
#>  [9,] -0.092112226 -0.116927112 -0.01581499 -0.07252787  0.0321061030
#> [10,] -0.086741559 -0.008206331 -0.08202893 -0.10321571 -0.1273860072
#>               V2          V20         V21          V22          V23         V24
#>  [1,] -0.1998724 -0.047342900 -0.01490930 -0.033685671  0.023023812  0.02954285
#>  [2,] -0.2280542  0.089052638  0.23064854  0.242282002  0.267362879  0.25366825
#>  [3,] -0.2345935  0.006097649  0.04365937  0.082434482 -0.004949119  0.12973051
#>  [4,] -0.2707332  0.220082874  0.20611399  0.134216412  0.094101753  0.17219678
#>  [5,] -0.1897159 -0.142394737  0.02212587 -0.006762076 -0.078903754 -0.02956018
#>  [6,] -0.4161573  0.076088557  0.14686275  0.040955867  0.222300322  0.19517047
#>  [7,] -0.2700188  0.058318280 -0.04136420 -0.079485723 -0.003189092  0.02895843
#>  [8,] -0.2951507  0.018436291  0.15110458  0.173129091  0.093168759  0.01684063
#>  [9,] -0.3318195  0.074793537  0.10197092  0.006262463  0.008106234  0.17653758
#> [10,] -0.3111658  0.079238770  0.04194581  0.151665643  0.143654080  0.12219665
#>               V25          V26         V27          V28         V29         V3
#>  [1,] -0.04757105  0.149388943  0.03765744  0.218180910  0.11417375 -0.3236960
#>  [2,]  0.23604675  0.281288380  0.28275598  0.176328627  0.14517524 -0.3034847
#>  [3,]  0.04176645  0.041791552  0.14568542  0.177223697  0.12268546 -0.1650339
#>  [4,]  0.30375327  0.127713572  0.22203015  0.206086337  0.22125694 -0.2832065
#>  [5,] -0.03581523  0.005403638  0.14236261 -0.006449771 -0.01001516 -0.1789689
#>  [6,]  0.20049280  0.157961176  0.22434673  0.194562075  0.18260851 -0.3581122
#>  [7,] -0.10807936 -0.088083168 -0.03742117  0.060633781  0.02271998 -0.2685821
#>  [8,]  0.17746978  0.029867130  0.14204150  0.132873423 -0.05044959 -0.3388977
#>  [9,]  0.08521756  0.055565745  0.13956618  0.117018591  0.17857427 -0.2661229
#> [10,]  0.19260851  0.221655689  0.08536572  0.268438872  0.25611462 -0.3169463
#>                V30         V31          V32          V33         V34
#>  [1,]  0.121361771  0.05793283  0.036874092 -0.007062410  0.01297698
#>  [2,]  0.188325607  0.04896377  0.008561258  0.150329753  0.07973101
#>  [3,]  0.150121161  0.10853970  0.123145567 -0.022249949  0.03184461
#>  [4,]  0.045374998 -0.01000196 -0.028300649  0.002127518 -0.03328855
#>  [5,]  0.007212666 -0.02642299 -0.114794649 -0.135515807 -0.12531646
#>  [6,]  0.116817423  0.06692692 -0.096874120 -0.030099814 -0.07964048
#>  [7,] -0.086819375 -0.13269565 -0.104668236 -0.012646900 -0.01457165
#>  [8,] -0.076201767 -0.18251779 -0.123181606 -0.171456142 -0.07835558
#>  [9,]  0.043796640  0.15018552  0.140226565  0.132348231  0.02826083
#> [10,]  0.189329780  0.01120095 -0.005586533 -0.064378244  0.09929240
#>                V35          V36          V37          V38          V39
#>  [1,]  0.068553034  0.037706606  0.056761059 -0.084721539  0.064050098
#>  [2,] -0.014855950  0.006216300 -0.084346398  0.063685016  0.075429785
#>  [3,] -0.075267423 -0.072035491  0.048422028  0.051141622  0.002784245
#>  [4,] -0.011983598  0.024699805  0.068257852  0.008295753  0.007813111
#>  [5,]  0.044370380  0.100606654 -0.077859097 -0.091961069 -0.095943427
#>  [6,] -0.073222226 -0.098928924 -0.039676509  0.022949347 -0.114360976
#>  [7,] -0.025242755 -0.008206255 -0.037649511 -0.147064864 -0.147124311
#>  [8,]  0.002526675 -0.131085140 -0.069931563 -0.148385253 -0.101938195
#>  [9,] -0.045479953  0.010713868 -0.042619938 -0.064175271 -0.032096570
#> [10,]  0.101331451  0.053342511  0.003414819  0.055464508  0.053093895
#>               V4          V40          V41         V42         V43         V44
#>  [1,] -0.2197498 -0.040096912  0.022307759  0.01815431 -0.14014494 -0.18990994
#>  [2,] -0.2757905  0.004775693 -0.049655883 -0.09256997 -0.04812464  0.03068287
#>  [3,] -0.1808167 -0.097024725 -0.047119823 -0.17104575 -0.06520180 -0.13469920
#>  [4,] -0.3183052 -0.046187769 -0.024142526 -0.02877414 -0.05745852 -0.07019854
#>  [5,] -0.2844433 -0.037372736 -0.106551382 -0.02732796 -0.04127912 -0.03399326
#>  [6,] -0.2704702 -0.090088387 -0.147873871 -0.12088968 -0.06642111 -0.03906867
#>  [7,] -0.1415618 -0.169007696 -0.082441353 -0.03742706 -0.15749998 -0.19333981
#>  [8,] -0.2597497 -0.072264938 -0.198266816 -0.10653330 -0.17770706 -0.20159639
#>  [9,] -0.2859873 -0.080175243 -0.009856653  0.01284588 -0.11633516 -0.07410122
#> [10,] -0.3336629  0.049222552 -0.034801609 -0.13302155 -0.04547742 -0.11947145
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.04128507 -0.12992105 -0.1141111 -0.1299916 -0.1716698 -0.1502174
#>  [2,] -0.05394076  0.02674948 -0.1442118 -0.2106451 -0.3078834 -0.1563003
#>  [3,] -0.22264041 -0.11404436 -0.1308951 -0.2044012 -0.3167660 -0.2277053
#>  [4,] -0.10535319 -0.15054652 -0.1310311 -0.2465057 -0.1742655 -0.1238278
#>  [5,] -0.18066236 -0.19340554 -0.1839966 -0.2087065 -0.2643226 -0.3053763
#>  [6,] -0.13925880 -0.17269152 -0.2825963 -0.3199832 -0.3461100 -0.3792521
#>  [7,] -0.15637537 -0.19399744 -0.1344274 -0.2694367 -0.3130584 -0.2106257
#>  [8,] -0.12416558 -0.24158736 -0.1777907 -0.2564189 -0.2969150 -0.1757067
#>  [9,] -0.02618976 -0.20826406 -0.2524056 -0.2007694 -0.2445196 -0.3294975
#> [10,] -0.07691486 -0.16226657 -0.2269080 -0.2657045 -0.2037367 -0.1717495
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3645655 -0.2853997 -0.2896127 -0.3112986 -0.2146496 -0.2728892
#>  [2,] -0.2404250 -0.3596087 -0.2270700 -0.3846038 -0.3718311 -0.3419289
#>  [3,] -0.3916353 -0.3084139 -0.4186664 -0.3335640 -0.3299330 -0.2706191
#>  [4,] -0.2926177 -0.3627439 -0.2105628 -0.2249420 -0.2454807 -0.4037006
#>  [5,] -0.1684265 -0.2977832 -0.3254349 -0.3251192 -0.1952184 -0.2816447
#>  [6,] -0.4399102 -0.4586328 -0.4560897 -0.4707370 -0.4520559 -0.3740876
#>  [7,] -0.3492721 -0.2030852 -0.2291764 -0.1954356 -0.3489723 -0.2283725
#>  [8,] -0.3983627 -0.2731917 -0.2885780 -0.3198165 -0.3786221 -0.3859037
#>  [9,] -0.3475998 -0.3070300 -0.4228137 -0.2653979 -0.4293721 -0.3874844
#> [10,] -0.3021986 -0.2703345 -0.3183848 -0.2793437 -0.3319129 -0.2750152
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.4030958 -0.3914998 -0.3194222 -0.3957356 -0.2031088 -0.3687045
#>  [2,] -0.4183517 -0.3208198 -0.3198351 -0.3810996 -0.2582415 -0.4129999
#>  [3,] -0.2488911 -0.3206142 -0.3443419 -0.2947128 -0.1682398 -0.3489686
#>  [4,] -0.2324560 -0.3161900 -0.3257272 -0.2355512 -0.1806202 -0.2396041
#>  [5,] -0.1874449 -0.2443099 -0.2253263 -0.2467002 -0.2163230 -0.2520033
#>  [6,] -0.4956586 -0.4150223 -0.4872661 -0.5254584 -0.1598978 -0.3722955
#>  [7,] -0.2947274 -0.3703323 -0.2285175 -0.3122091 -0.1719125 -0.3251824
#>  [8,] -0.3691869 -0.3051969 -0.3340110 -0.2744901 -0.1823451 -0.4392116
#>  [9,] -0.2991094 -0.3200055 -0.3376873 -0.3842528 -0.1970518 -0.4383304
#> [10,] -0.2795572 -0.3384751 -0.4165975 -0.2710641 -0.1313848 -0.3113142
#>               V7          V8          V9
#>  [1,] -0.1586252 -0.24997053 -0.06976396
#>  [2,] -0.1411307 -0.14135068 -0.13912783
#>  [3,] -0.1583093 -0.11197332 -0.13508365
#>  [4,] -0.1939920 -0.05168355 -0.05321277
#>  [5,] -0.1850253 -0.26637228 -0.14109685
#>  [6,] -0.2344111 -0.13802527 -0.15332323
#>  [7,] -0.2666469 -0.19716832 -0.20807756
#>  [8,] -0.2151243 -0.14375556 -0.12286941
#>  [9,] -0.1136068 -0.17976574 -0.04710346
#> [10,] -0.2392853 -0.09123069 -0.12247375
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,]  0.01116443  0.22682202  0.11612574  0.1301015 -0.05235683  0.19438369
#> [2,] -0.08036910 -0.06573539 -0.01337475 -0.1354683  0.09543187 -0.09693681
#>              [,7]        [,8]        [,9]       [,10]
#> [1,] -0.007663626  0.12397668  0.08697136  0.04235877
#> [2,] -0.048328801 -0.08388799 -0.06748276 -0.07287148
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12         V13
#>  [1,] 3.576897e-04 0.0027498473 0.0032127408 0.0033689748 0.003219027
#>  [2,] 1.498794e-04 0.0010912530 0.0012049313 0.0012539569 0.001308314
#>  [3,] 3.108944e-04 0.0023928494 0.0027607736 0.0029041087 0.002831934
#>  [4,] 1.778274e-04 0.0012564186 0.0013915238 0.0014476782 0.001486368
#>  [5,] 3.468585e-04 0.0028948892 0.0035134456 0.0036526590 0.003026740
#>  [6,] 1.476774e-04 0.0011573340 0.0012213928 0.0013005083 0.001467163
#>  [7,] 6.469433e-05 0.0005032995 0.0005788979 0.0006232028 0.000645603
#>  [8,] 1.107689e-04 0.0009117714 0.0009440374 0.0010440035 0.001277459
#>  [9,] 3.060980e-04 0.0023167236 0.0026660420 0.0028016755 0.002754762
#> [10,] 3.640388e-04 0.0026708165 0.0031211638 0.0032628077 0.003143124
#>                V14          V15          V16          V17         V18
#>  [1,] 0.0031065553 0.0031194136 0.0035878074 0.0039459369 0.004449152
#>  [2,] 0.0015014124 0.0017266049 0.0020363668 0.0022285454 0.002365663
#>  [3,] 0.0028519008 0.0029577698 0.0034351288 0.0037732439 0.004201744
#>  [4,] 0.0016531207 0.0018256762 0.0021253661 0.0023105171 0.002482793
#>  [5,] 0.0021769820 0.0017334753 0.0020232357 0.0023707268 0.002917080
#>  [6,] 0.0018889614 0.0023000801 0.0027761902 0.0030544614 0.003223975
#>  [7,] 0.0006859455 0.0007322924 0.0008629711 0.0009546396 0.001076063
#>  [8,] 0.0017377548 0.0021355249 0.0025659417 0.0027979176 0.002980959
#>  [9,] 0.0028313873 0.0029843864 0.0034580534 0.0037768338 0.004174126
#> [10,] 0.0030894377 0.0031367238 0.0035589879 0.0038640235 0.004302976
#>               V19           V2         V20         V21         V22         V23
#>  [1,] 0.005372797 4.299672e-04 0.006128714 0.006564987 0.006641979 0.006656792
#>  [2,] 0.002482691 1.928827e-04 0.002567617 0.002543592 0.002566614 0.002459869
#>  [3,] 0.004893731 3.803475e-04 0.005492761 0.005809138 0.005882957 0.005867962
#>  [4,] 0.002682385 2.228508e-04 0.002899420 0.002961721 0.003033241 0.002986813
#>  [5,] 0.004309809 3.870609e-04 0.005211616 0.005997035 0.006235384 0.006391306
#>  [6,] 0.003134109 2.119758e-04 0.003142667 0.003063720 0.003205122 0.003092753
#>  [7,] 0.001206927 8.302858e-05 0.001328260 0.001406910 0.001459199 0.001487235
#>  [8,] 0.002841949 1.672416e-04 0.002883858 0.002804069 0.002901543 0.002864197
#>  [9,] 0.004801742 3.780058e-04 0.005354212 0.005617263 0.005651365 0.005600041
#> [10,] 0.005171404 4.346442e-04 0.005896128 0.006246631 0.006209514 0.006168189
#>               V24         V25         V26         V27         V28         V29
#>  [1,] 0.006596962 0.006416381 0.006703150 0.006727626 0.006859857 0.006260687
#>  [2,] 0.002459902 0.002642841 0.002562252 0.002499375 0.002411680 0.002227208
#>  [3,] 0.005839586 0.005791759 0.005962620 0.005931303 0.005944295 0.005440175
#>  [4,] 0.003020782 0.003208005 0.003169372 0.003079946 0.003010268 0.002805514
#>  [5,] 0.006224293 0.005655179 0.006324308 0.006804089 0.007480781 0.006674016
#>  [6,] 0.003187418 0.003595272 0.003337828 0.003137261 0.002802907 0.002620595
#>  [7,] 0.001507254 0.001516020 0.001534442 0.001509024 0.001486010 0.001372311
#>  [8,] 0.002996963 0.003364189 0.003109134 0.002858347 0.002456111 0.002325992
#>  [9,] 0.005566921 0.005549484 0.005690573 0.005650741 0.005643142 0.005165358
#> [10,] 0.006068065 0.005874910 0.006142759 0.006155400 0.006319231 0.005794827
#>                 V3         V30         V31          V32          V33
#>  [1,] 4.983133e-04 0.005472802 0.004445745 0.0040039419 0.0036219530
#>  [2,] 2.316098e-04 0.002160325 0.002003337 0.0017437158 0.0016099727
#>  [3,] 4.470656e-04 0.004861127 0.004060918 0.0036207030 0.0033143412
#>  [4,] 2.659417e-04 0.002666683 0.002432736 0.0021627601 0.0020257154
#>  [5,] 4.300132e-04 0.005320925 0.003808989 0.0035290926 0.0027757040
#>  [6,] 2.608018e-04 0.002687097 0.002703146 0.0023017171 0.0021798696
#>  [7,] 9.716692e-05 0.001224758 0.001058380 0.0009229153 0.0008657011
#>  [8,] 2.213171e-04 0.002430872 0.002437424 0.0020119164 0.0020021676
#>  [9,] 4.437023e-04 0.004649309 0.003904492 0.0034611722 0.0031790779
#> [10,] 5.050927e-04 0.005137548 0.004189911 0.0037956013 0.0034665021
#>                V34          V35          V36          V37          V38
#>  [1,] 0.0032712479 0.0028496252 0.0024440361 0.0022632288 0.0028637449
#>  [2,] 0.0015418237 0.0016612223 0.0014712590 0.0013824689 0.0013631928
#>  [3,] 0.0030478761 0.0027832526 0.0024132731 0.0022236169 0.0026215622
#>  [4,] 0.0019047080 0.0019504334 0.0017224094 0.0015636448 0.0016007069
#>  [5,] 0.0021726921 0.0013367440 0.0009356213 0.0010284727 0.0022819437
#>  [6,] 0.0021343651 0.0024582483 0.0022855826 0.0021078363 0.0018068682
#>  [7,] 0.0008141486 0.0007686397 0.0007097227 0.0006566336 0.0006821865
#>  [8,] 0.0020076835 0.0022502506 0.0021071243 0.0018948198 0.0015182490
#>  [9,] 0.0029577784 0.0027602800 0.0024012386 0.0022268235 0.0025732524
#> [10,] 0.0031633003 0.0027542436 0.0022931923 0.0021071508 0.0027279137
#>                V39           V4          V40          V41          V42
#>  [1,] 0.0031420870 0.0007232606 0.0027579704 0.0028475205 0.0029752774
#>  [2,] 0.0012919711 0.0002727977 0.0012187303 0.0012248849 0.0012795311
#>  [3,] 0.0027923995 0.0006145903 0.0025153272 0.0025672772 0.0026839764
#>  [4,] 0.0015717282 0.0003242397 0.0014624368 0.0014977177 0.0015916594
#>  [5,] 0.0028863837 0.0007882080 0.0022482122 0.0025231171 0.0025265228
#>  [6,] 0.0015866693 0.0002281819 0.0016182563 0.0015328488 0.0015981558
#>  [7,] 0.0006827073 0.0001257430 0.0006442485 0.0006344578 0.0006232674
#>  [8,] 0.0013032460 0.0001675088 0.0014334095 0.0013224169 0.0013841315
#>  [9,] 0.0027000114 0.0006006022 0.0024313103 0.0024701900 0.0025969694
#> [10,] 0.0029978487 0.0007376616 0.0026135019 0.0027419029 0.0029050107
#>                V43          V44          V45          V46          V47
#>  [1,] 0.0027453251 0.0026246884 0.0026069857 0.0020925084 0.0015423573
#>  [2,] 0.0010254853 0.0008859636 0.0009210232 0.0007526018 0.0005803911
#>  [3,] 0.0023903682 0.0022405645 0.0022476150 0.0018093942 0.0013484595
#>  [4,] 0.0013023042 0.0011497353 0.0011913928 0.0009873496 0.0007458627
#>  [5,] 0.0027373687 0.0029640431 0.0029074361 0.0023279785 0.0016792233
#>  [6,] 0.0011310663 0.0008622237 0.0008961810 0.0007413299 0.0006277513
#>  [7,] 0.0005361074 0.0004768797 0.0004563404 0.0003762977 0.0002998701
#>  [8,] 0.0009339880 0.0006448315 0.0006706425 0.0005593018 0.0005072125
#>  [9,] 0.0022933132 0.0021256183 0.0021412464 0.0017273045 0.0012868486
#> [10,] 0.0026563359 0.0025148197 0.0025301011 0.0020370352 0.0014790548
#>                V48          V49           V5          V50          V51
#>  [1,] 0.0012306289 0.0006844874 0.0009493600 2.233337e-04 2.146558e-04
#>  [2,] 0.0004261277 0.0002135416 0.0003799002 8.506181e-05 7.715909e-05
#>  [3,] 0.0010593025 0.0005790592 0.0008257382 1.964593e-04 1.834189e-04
#>  [4,] 0.0005517600 0.0002811950 0.0004369153 1.046525e-04 9.599481e-05
#>  [5,] 0.0013646813 0.0007937499 0.0009667851 2.197202e-04 2.264773e-04
#>  [6,] 0.0004317632 0.0001999488 0.0003890330 9.410539e-05 7.502587e-05
#>  [7,] 0.0002308435 0.0001273888 0.0001770559 4.604947e-05 3.952839e-05
#>  [8,] 0.0003510966 0.0001570339 0.0003142251 8.270444e-05 5.793517e-05
#>  [9,] 0.0010089706 0.0005469064 0.0008055477 1.889853e-04 1.770421e-04
#> [10,] 0.0011844777 0.0006552221 0.0009347016 2.160787e-04 2.110737e-04
#>                V52          V53          V54          V55          V56
#>  [1,] 1.852428e-04 1.269261e-04 1.288515e-04 9.394003e-05 8.365078e-05
#>  [2,] 7.672138e-05 5.356080e-05 4.680763e-05 4.937356e-05 3.609963e-05
#>  [3,] 1.619015e-04 1.112155e-04 1.114306e-04 8.765139e-05 7.461603e-05
#>  [4,] 9.186844e-05 6.256466e-05 5.680778e-05 5.576943e-05 4.146938e-05
#>  [5,] 1.852621e-04 1.140111e-04 1.215075e-04 5.863349e-05 7.136356e-05
#>  [6,] 7.952387e-05 5.669340e-05 4.974310e-05 6.438246e-05 4.169174e-05
#>  [7,] 3.362683e-05 2.474326e-05 2.534782e-05 2.077858e-05 1.794098e-05
#>  [8,] 5.949440e-05 4.570484e-05 4.421637e-05 5.619938e-05 3.736227e-05
#>  [9,] 1.579994e-04 1.093950e-04 1.086896e-04 8.771592e-05 7.344754e-05
#> [10,] 1.855727e-04 1.273422e-04 1.271336e-04 9.546926e-05 8.284090e-05
#>                V57          V58          V59           V6          V60
#>  [1,] 6.919650e-05 9.228732e-05 8.705093e-05 0.0011189149 7.028825e-05
#>  [2,] 3.257737e-05 3.497142e-05 3.675232e-05 0.0004855038 3.777651e-05
#>  [3,] 6.287481e-05 7.847442e-05 7.632090e-05 0.0010069262 6.543675e-05
#>  [4,] 3.729219e-05 4.046501e-05 4.435567e-05 0.0005509009 4.304496e-05
#>  [5,] 5.732972e-05 9.266692e-05 7.677619e-05 0.0009876407 4.783223e-05
#>  [6,] 4.003977e-05 3.299101e-05 3.882314e-05 0.0006053461 4.673934e-05
#>  [7,] 1.529919e-05 1.734300e-05 1.698974e-05 0.0002373376 1.511808e-05
#>  [8,] 3.490754e-05 2.520426e-05 3.230473e-05 0.0005460485 4.135305e-05
#>  [9,] 6.214030e-05 7.655322e-05 7.487016e-05 0.0009822284 6.586479e-05
#> [10,] 6.851164e-05 9.236804e-05 8.788846e-05 0.0010744034 7.116933e-05
#>                 V7           V8           V9
#>  [1,] 0.0012186156 0.0014170531 0.0022831118
#>  [2,] 0.0005009931 0.0005329934 0.0008491976
#>  [3,] 0.0010844464 0.0012400773 0.0019779786
#>  [4,] 0.0005737498 0.0006254874 0.0010027516
#>  [5,] 0.0010966615 0.0013541200 0.0024390743
#>  [6,] 0.0006033305 0.0006264506 0.0008896408
#>  [7,] 0.0002664129 0.0002941377 0.0004190424
#>  [8,] 0.0005531344 0.0005641907 0.0007290332
#>  [9,] 0.0010545156 0.0011984226 0.0019050241
#> [10,] 0.0011678373 0.0013552305 0.0022098946
#> 
#> $vW[[2]]
#>            [,1]       [,2]        [,3]      [,4]        [,5]        [,6]
#> [1,] -0.0690805 -0.1028172 -0.07094863 -0.097979 -0.03045681 -0.08462537
#> [2,]  0.0690805  0.1028172  0.07094863  0.097979  0.03045681  0.08462537
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.02174563 -0.05656284 -0.07912543 -0.08304549
#> [2,]  0.02174563  0.05656284  0.07912543  0.08304549
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3875034 0.7084103 0.3004867 0.5563674 0.1791375 0.6609423 0.0768114
#>  [8] 0.4743246 0.5563068 0.3132047
#> 
#> $B[[2]]
#> [1]  0.1903559 -0.2030972
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.009513207 0.004087189 0.008561147 0.004811206 0.008195037 0.005192630
#>  [7] 0.002203727 0.004708613 0.008307800 0.009056188
#> 
#> $vB[[2]]
#> [1] -0.1151768  0.1151768
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#>  [2,] 0.0459 0.1408 0.2693 0.3259 0.4545 0.5785 0.4471 0.2231 0.2164 0.3201
#>  [3,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#>  [4,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#>  [5,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#>  [6,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [7,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#>  [8,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#>  [9,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [10,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [11,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [12,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [13,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [14,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [15,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [16,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [17,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [18,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [19,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [20,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [21,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [22,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [23,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [24,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [25,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [26,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [27,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [28,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [29,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [30,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [31,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [32,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [33,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [34,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [35,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [36,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [37,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [38,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [39,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [40,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#>  [2,] 0.2915 0.0437 0.4235 0.4460 0.2380 0.6415 0.8966 0.8918 0.7529 0.6838
#>  [3,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#>  [4,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#>  [5,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#>  [6,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [7,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#>  [8,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#>  [9,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [10,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [11,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [12,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [13,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [14,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [15,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [16,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [17,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [18,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [19,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [20,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [21,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [22,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [23,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [24,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [25,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [26,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [27,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [28,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [29,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [30,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [31,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [32,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [33,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [34,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [35,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [36,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [37,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [38,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [39,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [40,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#>  [2,] 0.8390 1.0000 0.0347 0.8362 0.5427 0.4577 0.8067 0.6973 0.3915 0.1558
#>  [3,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#>  [4,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#>  [5,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#>  [6,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [7,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#>  [8,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#>  [9,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [10,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [11,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [12,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [13,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [14,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [15,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [16,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [17,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [18,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [19,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [20,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [21,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [22,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [23,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [24,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [25,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [26,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [27,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [28,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [29,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [30,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [31,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [32,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [33,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [34,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [35,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [36,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [37,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [38,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [39,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [40,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#>  [2,] 0.1598 0.2161 0.5178 0.0456 0.4782 0.2344 0.3599 0.2785 0.1807 0.0352
#>  [3,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#>  [4,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#>  [5,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#>  [6,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [7,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#>  [8,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#>  [9,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [10,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [11,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [12,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [13,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [14,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [15,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [16,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [17,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [18,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [19,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [20,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [21,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [22,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [23,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [24,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [25,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [26,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [27,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [28,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [29,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [30,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [31,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [32,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [33,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [34,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [35,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [36,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [37,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [38,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [39,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [40,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#>  [2,] 0.0473 0.0322 0.0408 0.0163 0.0067 0.0088 0.0121 0.0067 0.0032 0.0109
#>  [3,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#>  [4,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#>  [5,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#>  [6,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [7,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#>  [8,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#>  [9,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [10,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [11,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [12,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [13,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [14,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [15,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [16,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [17,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [18,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [19,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [20,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [21,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [22,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [23,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [24,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [25,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [26,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [27,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [28,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [29,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [30,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [31,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [32,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [33,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [34,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [35,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [36,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [37,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [38,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [39,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [40,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#>  [2,] 0.0164 0.0151 0.0070 0.0085 0.0117 0.0890 0.0056 0.1798 0.1741 0.1598
#>  [3,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#>  [4,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#>  [5,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#>  [6,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [7,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#>  [8,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#>  [9,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [10,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [11,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [12,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [13,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [14,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [15,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [16,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [17,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [18,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [19,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [20,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [21,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [22,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [23,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [24,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [25,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [26,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [27,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [28,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [29,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [30,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [31,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [32,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [33,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [34,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [35,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [36,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [37,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [38,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [39,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [40,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5481698 0.8690302 0.5612309 0.8641073 0.2210145 0.6968212 0.1436333
#>  [2,] 0.6398646 0.9087226 0.6455853 0.8619765 0.2854723 0.7846783 0.2346720
#>  [3,] 0.5937843 0.9195464 0.6462713 0.9057766 0.3100707 0.8340429 0.2490037
#>  [4,] 0.6792965 0.9206971 0.7167461 0.8676246 0.3711887 0.8268712 0.2862317
#>  [5,] 0.6113683 0.8737504 0.6040324 0.8479676 0.3397864 0.7352264 0.2627962
#>  [6,] 0.5812607 0.8739867 0.5712236 0.8629046 0.2741905 0.7249314 0.1812934
#>  [7,] 0.6124358 0.9198186 0.6187511 0.9041117 0.2942710 0.8087286 0.2348011
#>  [8,] 0.4835230 0.8462228 0.5112952 0.8518579 0.1819140 0.6842388 0.1647079
#>  [9,] 0.5566515 0.8909495 0.5662377 0.8662077 0.2441990 0.7296780 0.2035276
#> [10,] 0.6540052 0.9345693 0.6963019 0.9061589 0.3480235 0.8529748 0.2814738
#> [11,] 0.6814697 0.9032712 0.7020590 0.8521235 0.3573370 0.7941840 0.2766244
#> [12,] 0.5449516 0.8410105 0.5646512 0.8074507 0.3047267 0.7070084 0.2404806
#> [13,] 0.6244653 0.8948758 0.6559964 0.8719079 0.2924927 0.7757796 0.2371723
#> [14,] 0.5558327 0.7983180 0.5469287 0.7554316 0.3003189 0.6444584 0.2412331
#> [15,] 0.5607741 0.8762603 0.5750558 0.8769562 0.2244145 0.7293011 0.1906955
#> [16,] 0.6059244 0.9257090 0.6610726 0.9066193 0.3036066 0.8362592 0.2459472
#> [17,] 0.6769510 0.9256468 0.6661656 0.8962052 0.3530391 0.8107043 0.2522711
#> [18,] 0.6360174 0.8759331 0.5996340 0.7932318 0.2580270 0.6516328 0.1636103
#> [19,] 0.6301912 0.8900857 0.6594575 0.8381997 0.3333640 0.7654111 0.2374824
#> [20,] 0.5232965 0.7715284 0.4797000 0.7538466 0.2110505 0.4940967 0.1661107
#> [21,] 0.5789227 0.9103428 0.6045245 0.8738091 0.2465830 0.7648034 0.1934270
#> [22,] 0.5674445 0.8919128 0.5964451 0.8697980 0.3046503 0.7630529 0.2490208
#> [23,] 0.6205656 0.9105369 0.6195997 0.8638321 0.3265496 0.7847753 0.2626860
#> [24,] 0.6062045 0.8981480 0.6381777 0.8984442 0.2663250 0.8014796 0.2283479
#> [25,] 0.5359212 0.8805316 0.5411772 0.8726905 0.2394036 0.7284538 0.1919587
#> [26,] 0.6496753 0.9340080 0.6869302 0.9108854 0.2973394 0.8418758 0.2257819
#> [27,] 0.6457243 0.8514431 0.6154304 0.7757035 0.3321467 0.6447783 0.2211429
#> [28,] 0.6523445 0.8870004 0.6288320 0.8164596 0.2928495 0.7076918 0.1910897
#> [29,] 0.5892751 0.8997540 0.6137884 0.8930621 0.2252042 0.7805170 0.1886595
#> [30,] 0.4795542 0.8841974 0.5190774 0.8836998 0.1954098 0.7382863 0.1443690
#> [31,] 0.6022309 0.9201336 0.6482554 0.9059134 0.2788760 0.8226724 0.2111392
#> [32,] 0.5341400 0.8799992 0.5497316 0.8706232 0.2381500 0.7265436 0.2047741
#> [33,] 0.5571005 0.8508667 0.5341779 0.7820366 0.2170514 0.6278151 0.1466558
#> [34,] 0.6613885 0.8930381 0.6734985 0.8555000 0.3543033 0.7959627 0.2856727
#> [35,] 0.5845858 0.9156692 0.6144821 0.8910437 0.2368878 0.7584524 0.2016429
#> [36,] 0.6362269 0.9018394 0.6770742 0.8570119 0.3165928 0.7919626 0.2564549
#> [37,] 0.5125276 0.8245508 0.5457945 0.8508690 0.1598096 0.6314749 0.1244346
#> [38,] 0.6437330 0.9326240 0.6893244 0.8990613 0.3485456 0.8475131 0.2747548
#> [39,] 0.6196682 0.9138813 0.6232095 0.8663188 0.3110904 0.7764587 0.2423613
#> [40,] 0.5571138 0.8577908 0.5527376 0.8579784 0.2719062 0.7217308 0.2030398
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5059610 0.6313744 0.6741196
#>  [2,] 0.5012712 0.7363020 0.7742395
#>  [3,] 0.7001227 0.7181005 0.7394400
#>  [4,] 0.5778592 0.7763902 0.7971331
#>  [5,] 0.5520723 0.6828912 0.7061303
#>  [6,] 0.5270678 0.6462924 0.7000692
#>  [7,] 0.6268329 0.7020206 0.7662856
#>  [8,] 0.5288067 0.5567698 0.5895822
#>  [9,] 0.5381668 0.6658331 0.6926781
#> [10,] 0.6539499 0.7620724 0.8014418
#> [11,] 0.5039000 0.7738170 0.7952130
#> [12,] 0.5430730 0.6122660 0.6374547
#> [13,] 0.5497586 0.7084224 0.7283495
#> [14,] 0.4588190 0.6252082 0.6142727
#> [15,] 0.5805492 0.6434369 0.6507047
#> [16,] 0.6866732 0.7349974 0.7543426
#> [17,] 0.5581865 0.7351159 0.8189982
#> [18,] 0.3006192 0.6929036 0.7280433
#> [19,] 0.5068347 0.7180470 0.7468025
#> [20,] 0.3137092 0.5608543 0.6160507
#> [21,] 0.5171167 0.6877827 0.7313950
#> [22,] 0.6013402 0.6653638 0.7088102
#> [23,] 0.5466249 0.7260009 0.7600862
#> [24,] 0.6300837 0.6871049 0.7286593
#> [25,] 0.5594920 0.6340141 0.6637847
#> [26,] 0.6139659 0.7533834 0.7963601
#> [27,] 0.3671805 0.6939880 0.7331165
#> [28,] 0.3647139 0.7154052 0.7550147
#> [29,] 0.5751256 0.6806174 0.7079238
#> [30,] 0.5832192 0.5887026 0.6488157
#> [31,] 0.6280010 0.7128316 0.7568643
#> [32,] 0.5685078 0.6269330 0.6666064
#> [33,] 0.3310450 0.6394204 0.6383335
#> [34,] 0.5676737 0.7335219 0.7605509
#> [35,] 0.5260273 0.7037148 0.7376736
#> [36,] 0.5554040 0.7333378 0.7551954
#> [37,] 0.4071080 0.5778839 0.5993461
#> [38,] 0.6385753 0.7544341 0.7951592
#> [39,] 0.5232489 0.7233611 0.7638370
#> [40,] 0.5677523 0.6307599 0.6574259
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5783619 0.4216381
#>  [2,] 0.5756920 0.4243080
#>  [3,] 0.5824339 0.4175661
#>  [4,] 0.5736799 0.4263201
#>  [5,] 0.5729678 0.4270322
#>  [6,] 0.5762952 0.4237048
#>  [7,] 0.5803357 0.4196643
#>  [8,] 0.5817990 0.4182010
#>  [9,] 0.5789893 0.4210107
#> [10,] 0.5786614 0.4213386
#> [11,] 0.5711138 0.4288862
#> [12,] 0.5741327 0.4258673
#> [13,] 0.5768035 0.4231965
#> [14,] 0.5689438 0.4310562
#> [15,] 0.5813092 0.4186908
#> [16,] 0.5822823 0.4177177
#> [17,] 0.5742295 0.4257705
#> [18,] 0.5674413 0.4325587
#> [19,] 0.5724255 0.4275745
#> [20,] 0.5657051 0.4342949
#> [21,] 0.5791048 0.4208952
#> [22,] 0.5780021 0.4219979
#> [23,] 0.5751849 0.4248151
#> [24,] 0.5815568 0.4184432
#> [25,] 0.5802850 0.4197150
#> [26,] 0.5802410 0.4197590
#> [27,] 0.5638981 0.4361019
#> [28,] 0.5688076 0.4311924
#> [29,] 0.5822537 0.4177463
#> [30,] 0.5845860 0.4154140
#> [31,] 0.5817461 0.4182539
#> [32,] 0.5804108 0.4195892
#> [33,] 0.5710883 0.4289117
#> [34,] 0.5734192 0.4265808
#> [35,] 0.5798272 0.4201728
#> [36,] 0.5755152 0.4244848
#> [37,] 0.5775222 0.4224778
#> [38,] 0.5781829 0.4218171
#> [39,] 0.5752665 0.4247335
#> [40,] 0.5776891 0.4223109
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]     [,2]        [,3]     [,4]       [,5]        [,6]
#>  [1,]  0.19327873 1.892411  0.24615919 1.849831 -1.2597644  0.83220623
#>  [2,]  0.57477663 2.298137  0.59968985 1.831804 -0.9174769  1.29314041
#>  [3,]  0.37963175 2.436201  0.60268902 2.263124 -0.7997889  1.61455535
#>  [4,]  0.75054066 2.451856  0.92837803 1.880117 -0.5271209  1.56361296
#>  [5,]  0.45306734 1.934534  0.42229528 1.718749 -0.6642461  1.02130327
#>  [6,]  0.32795096 1.936677  0.28684505 1.839627 -0.9734646  0.96905673
#>  [7,]  0.45756248 2.439884  0.48425078 2.243770 -0.8747302  1.44176981
#>  [8,] -0.06593194 1.705278  0.04518867 1.749248 -1.5034336  0.77332056
#>  [9,]  0.22758330 2.100477  0.26651736 1.867836 -1.1297947  0.99298956
#> [10,]  0.63669143 2.659094  0.82974944 2.267612 -0.6277385  1.75812633
#> [11,]  0.76053402 2.234112  0.85712210 1.751354 -0.5869413  1.35033249
#> [12,]  0.18029311 1.665766  0.26006056 1.433530 -0.8248898  0.88089869
#> [13,]  0.50854498 2.141542  0.64550306 1.917934 -0.8833084  1.24123901
#> [14,]  0.22426607 1.375815  0.18826897 1.127794 -0.8457798  0.59476813
#> [15,]  0.24430420 1.957483  0.30250913 1.963916 -1.2401234  0.99107946
#> [16,]  0.43021246 2.522570  0.66807796 2.273038 -0.8301819  1.63065416
#> [17,]  0.73979510 2.521666  0.69089327 2.155753 -0.6057070  1.45459309
#> [18,]  0.55811975 1.954469  0.40394016 1.344517 -1.0562484  0.62622423
#> [19,]  0.53303700 2.091617  0.66087757 1.644893 -0.6930094  1.18257859
#> [20,]  0.09325350 1.216962 -0.08124453 1.119234 -1.3186052 -0.02361434
#> [21,]  0.31835267 2.317828  0.42435337 1.935067 -1.1169200  1.17919686
#> [22,]  0.27143241 2.110430  0.39067446 1.899174 -0.8252504  1.16949045
#> [23,]  0.49194974 2.320208  0.48784943 1.847490 -0.7238324  1.29371482
#> [24,]  0.43138573 2.176814  0.56746349 2.180057 -1.0133485  1.39556744
#> [25,]  0.14393262 1.997474  0.16508252 1.924960 -1.1559520  0.98679204
#> [26,]  0.61761248 2.649952  0.78580633 2.324494 -0.8599995  1.67225149
#> [27,]  0.60029732 1.745964  0.47019721 1.240801 -0.6984917  0.59616433
#> [28,]  0.62936073 2.060461  0.52720915 1.492543 -0.8815848  0.88420002
#> [29,]  0.36096948 2.194495  0.46326464 2.122408 -1.2355919  1.26868172
#> [30,] -0.08182887 2.032793  0.07634678 2.027943 -1.4152344  1.03708063
#> [31,]  0.41476914 2.444163  0.61137958 2.264728 -0.9500440  1.53455945
#> [32,]  0.13677278 1.992422  0.19958621 1.906480 -1.1628490  0.97715607
#> [33,]  0.22940261 1.741415  0.13692518 1.277575 -1.2829331  0.52285493
#> [34,]  0.66948778 2.122156  0.72405110 1.778407 -0.6001764  1.36124926
#> [35,]  0.34162741 2.384908  0.46619203 2.101447 -1.1698182  1.14421351
#> [36,]  0.55902504 2.217831  0.74035847 1.790690 -0.7694743  1.33679600
#> [37,]  0.05012105 1.547489  0.18369264 1.741433 -1.6596456  0.53854938
#> [38,]  0.59160353 2.627713  0.79696253 2.186838 -0.6254383  1.71522791
#> [39,]  0.48814002 2.361974  0.50319325 1.868795 -0.7950267  1.24514746
#> [40,]  0.22945660 1.797061  0.21173792 1.798600 -0.9849728  0.95306318
#>             [,7]         [,8]      [,9]     [,10]
#>  [1,] -1.7854352  0.023845027 0.5381175 0.7268769
#>  [2,] -1.1821155  0.005084745 1.0268357 1.2324067
#>  [3,] -1.1039328  0.847882299 0.9350589 1.0430598
#>  [4,] -0.9137570  0.313991417 1.2447527 1.3684716
#>  [5,] -1.0314853  0.209047339 0.7670906 0.8766631
#>  [6,] -1.5076091  0.108377184 0.6027815 0.8476272
#>  [7,] -1.1813972  0.518653719 0.8569383 1.1874550
#>  [8,] -1.6236077  0.115354602 0.2280624 0.3622384
#>  [9,] -1.3643910  0.152964772 0.6893984 0.8126698
#> [10,] -0.9371630  0.636446984 1.1640752 1.3953302
#> [11,] -0.9612680  0.015600400 1.2299911 1.3566396
#> [12,] -1.1500464  0.172720073 0.4568471 0.5643336
#> [13,] -1.1682455  0.199695554 0.8877343 0.9862647
#> [14,] -1.1459310 -0.165097803 0.5117141 0.4653081
#> [15,] -1.4454975  0.325028282 0.5903128 0.6221381
#> [16,] -1.1203454  0.784611570 1.0201273 1.1219088
#> [17,] -1.0865360  0.233805439 1.0207359 1.5095751
#> [18,] -1.6316070 -0.844350839 0.8137290 0.9847178
#> [19,] -1.1665320  0.027340640 0.9347944 1.0816307
#> [20,] -1.6134465 -0.782834943 0.2446300 0.4728188
#> [21,] -1.4278945  0.068493446 0.7897735 1.0017120
#> [22,] -1.1038417  0.411052343 0.6872902 0.8896125
#> [23,] -1.0320546  0.187042849 0.9744266 1.1531523
#> [24,] -1.2176636  0.532575927 0.7866189 0.9878310
#> [25,] -1.4373326  0.239100661 0.5494765 0.6802063
#> [26,] -1.2322840  0.464013812 1.1167395 1.3636983
#> [27,] -1.2590184 -0.544331828 0.8188304 1.0104922
#> [28,] -1.4429451 -0.554962176 0.9217828 1.1255392
#> [29,] -1.4587443  0.302794913 0.7566108 0.8853218
#> [30,] -1.7794669  0.336002473 0.3586047 0.6138373
#> [31,] -1.3180723  0.523650580 0.9091764 1.1355641
#> [32,] -1.3567189  0.275765629 0.5190818 0.6928760
#> [33,] -1.7610750 -0.703462283 0.5728494 0.5681383
#> [34,] -0.9164943  0.272366233 1.0125653 1.1557021
#> [35,] -1.3760579  0.104203304 0.8650508 1.0339118
#> [36,] -1.0644765  0.222529936 1.0116239 1.1265163
#> [37,] -1.9510899 -0.375933667 0.3140927 0.4027411
#> [38,] -0.9706306  0.569185697 1.1224028 1.3563091
#> [39,] -1.1397768  0.093062612 0.9611956 1.1738320
#> [40,] -1.3674030  0.272686517 0.5354783 0.6518440
#> 
#> $pre[[3]]
#>            [,1]        [,2]
#>  [1,] 0.2698968 -0.04615571
#>  [2,] 0.2713119 -0.03380128
#>  [3,] 0.2943989 -0.03837396
#>  [4,] 0.2744717 -0.02240960
#>  [5,] 0.2613223 -0.03264782
#>  [6,] 0.2670659 -0.04051706
#>  [7,] 0.2848441 -0.03930767
#>  [8,] 0.2737948 -0.05636802
#>  [9,] 0.2740286 -0.04459718
#> [10,] 0.2877362 -0.02954453
#> [11,] 0.2639375 -0.02245928
#> [12,] 0.2601243 -0.03860857
#> [13,] 0.2745615 -0.03510336
#> [14,] 0.2405708 -0.03697245
#> [15,] 0.2790756 -0.04907446
#> [16,] 0.2946437 -0.03750584
#> [17,] 0.2720420 -0.02708669
#> [18,] 0.2409567 -0.03046260
#> [19,] 0.2646087 -0.02714531
#> [20,] 0.2187838 -0.04556550
#> [21,] 0.2782503 -0.04084926
#> [22,] 0.2759207 -0.03865641
#> [23,] 0.2708778 -0.03215975
#> [24,] 0.2865548 -0.04261256
#> [25,] 0.2759799 -0.04796356
#> [26,] 0.2897059 -0.03405697
#> [27,] 0.2332150 -0.02378247
#> [28,] 0.2493098 -0.02767819
#> [29,] 0.2858193 -0.04621259
#> [30,] 0.2870295 -0.05459890
#> [31,] 0.2909132 -0.03903227
#> [32,] 0.2762918 -0.04816824
#> [33,] 0.2456614 -0.04063134
#> [34,] 0.2682235 -0.02759173
#> [35,] 0.2799207 -0.04214333
#> [36,] 0.2739147 -0.03047487
#> [37,] 0.2597422 -0.05286766
#> [38,] 0.2862608 -0.02905751
#> [39,] 0.2708034 -0.03256829
#> [40,] 0.2690169 -0.04427733
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5783619  0.5783619
#>  [2,] -0.5756920  0.5756920
#>  [3,] -0.5824339  0.5824339
#>  [4,]  0.4263201 -0.4263201
#>  [5,] -0.5729678  0.5729678
#>  [6,] -0.5762952  0.5762952
#>  [7,]  0.4196643 -0.4196643
#>  [8,]  0.4182010 -0.4182010
#>  [9,]  0.4210107 -0.4210107
#> [10,]  0.4213386 -0.4213386
#> [11,] -0.5711138  0.5711138
#> [12,]  0.4258673 -0.4258673
#> [13,]  0.4231965 -0.4231965
#> [14,]  0.4310562 -0.4310562
#> [15,] -0.5813092  0.5813092
#> [16,] -0.5822823  0.5822823
#> [17,]  0.4257705 -0.4257705
#> [18,]  0.4325587 -0.4325587
#> [19,]  0.4275745 -0.4275745
#> [20,]  0.4342949 -0.4342949
#> [21,]  0.4208952 -0.4208952
#> [22,] -0.5780021  0.5780021
#> [23,] -0.5751849  0.5751849
#> [24,] -0.5815568  0.5815568
#> [25,]  0.4197150 -0.4197150
#> [26,]  0.4197590 -0.4197590
#> [27,] -0.5638981  0.5638981
#> [28,]  0.4311924 -0.4311924
#> [29,] -0.5822537  0.5822537
#> [30,]  0.4154140 -0.4154140
#> [31,]  0.4182539 -0.4182539
#> [32,]  0.4195892 -0.4195892
#> [33,]  0.4289117 -0.4289117
#> [34,]  0.4265808 -0.4265808
#> [35,]  0.4201728 -0.4201728
#> [36,]  0.4244848 -0.4244848
#> [37,] -0.5775222  0.5775222
#> [38,]  0.4218171 -0.4218171
#> [39,] -0.5752665  0.5752665
#> [40,] -0.5776891  0.5776891
#> 
#> $L
#> [1] 0.6914657 0.6880063 0.7121351 0.7445590 1.0419696 0.6750767
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609