Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11          V12         V13
#>  [1,] -0.3003702 -0.138904512 -0.029265763  0.009706611 -0.04739455
#>  [2,] -0.3497380  0.005735713  0.002926425 -0.060534614 -0.10558186
#>  [3,] -0.3286378 -0.166004052 -0.084699347 -0.114898071 -0.04880862
#>  [4,] -0.2094940 -0.091685154 -0.001057832 -0.153678790  0.02701012
#>  [5,] -0.3685980  0.004656026 -0.109713640 -0.042080793 -0.08041415
#>  [6,] -0.2317832 -0.196032775 -0.179077179 -0.055421604 -0.02183681
#>  [7,] -0.2266461 -0.145493695 -0.058693450 -0.079239037 -0.19880188
#>  [8,] -0.3041006 -0.192289112 -0.172207090 -0.089462552 -0.01697174
#>  [9,] -0.2336526 -0.131709337 -0.245079086 -0.186193073 -0.18406560
#> [10,] -0.4089008 -0.107736258 -0.110133037 -0.070050002 -0.12059124
#>                V14          V15         V16          V17          V18
#>  [1,] -0.101482900 -0.107974522 -0.11904725  0.083604716  0.052828504
#>  [2,] -0.128255342  0.047151881  0.07197774  0.078366113  0.092167781
#>  [3,]  0.004390151 -0.133420093 -0.15274927 -0.119121127 -0.014698182
#>  [4,] -0.068978164  0.009743125  0.07785396 -0.042770628  0.015327764
#>  [5,]  0.009085303 -0.085758638 -0.10041445  0.008824263 -0.104052300
#>  [6,]  0.047007035 -0.079639692 -0.05957572 -0.030505869  0.022833929
#>  [7,] -0.002773052 -0.052787677 -0.10598087 -0.009887933 -0.046755951
#>  [8,] -0.176421627 -0.160296702 -0.04771174 -0.136406086 -0.006648575
#>  [9,] -0.140319887 -0.190203657 -0.15834006 -0.111149187 -0.114579416
#> [10,]  0.017459609 -0.020313081  0.05020163 -0.139357780  0.037892642
#>               V19         V2          V20         V21        V22         V23
#>  [1,]  0.09286089 -0.1728791  0.183228625  0.07776273 0.14883246  0.11173990
#>  [2,]  0.03633567 -0.2928786  0.188026118  0.20114925 0.25334906  0.23970982
#>  [3,]  0.05553978 -0.3061983 -0.045586502  0.07240301 0.07912382  0.18851660
#>  [4,]  0.06868150 -0.3038524  0.168712557  0.21013177 0.21766955  0.27757468
#>  [5,]  0.13153989 -0.2667465  0.098650008  0.17915780 0.14406137  0.02838727
#>  [6,] -0.04181684 -0.3590625 -0.073771933  0.02732194 0.14573104  0.11992967
#>  [7,]  0.02443857 -0.2625517 -0.100054381  0.04240211 0.07345167  0.02411071
#>  [8,] -0.06298968 -0.2735286  0.085078088  0.08705953 0.15203805  0.12393524
#>  [9,] -0.14840139 -0.2319341 -0.001416584 -0.03838981 0.06773017 -0.02356917
#> [10,] -0.01981329 -0.2849177  0.113711565  0.17291672 0.04719213  0.05967327
#>              V24        V25        V26       V27        V28          V29
#>  [1,] 0.11996757 0.15677876 0.24836919 0.2323333 0.23398444  0.164992395
#>  [2,] 0.30281282 0.24252103 0.25349779 0.1960707 0.15012340  0.166278434
#>  [3,] 0.02436295 0.07165134 0.09909212 0.1257131 0.11614428 -0.006281422
#>  [4,] 0.26589469 0.29739479 0.33164428 0.2330838 0.27562058  0.135323953
#>  [5,] 0.11510785 0.01700621 0.06579918 0.1951628 0.21992951  0.171951930
#>  [6,] 0.05818401 0.18311907 0.20232692 0.2387354 0.14108464  0.159128423
#>  [7,] 0.16519534 0.16215383 0.05704942 0.1392059 0.08461018  0.062481291
#>  [8,] 0.06947529 0.15220352 0.21499581 0.1581357 0.10125187  0.163286127
#>  [9,] 0.01113038 0.08503001 0.06888085 0.0562689 0.04191849  0.057816167
#> [10,] 0.15993364 0.16030219 0.27570143 0.2466372 0.18593492  0.093380488
#>               V3          V30          V31          V32         V33         V34
#>  [1,] -0.2294890  0.194951750  0.121646861  0.146016843  0.13624920  0.14531051
#>  [2,] -0.3429622  0.027772104  0.069221029  0.087744158  0.06736024  0.01197792
#>  [3,] -0.2924057  0.007969888  0.048669487 -0.130502960 -0.13824222 -0.02817686
#>  [4,] -0.2077880  0.143162206  0.150440063 -0.008860528 -0.05964300  0.03663829
#>  [5,] -0.3524318  0.071630008 -0.043201145  0.094001618 -0.03361987 -0.00216898
#>  [6,] -0.2247324 -0.032411360 -0.014455904 -0.001636120 -0.10083944 -0.08454841
#>  [7,] -0.2172097  0.027007659  0.071357205 -0.057688075 -0.08565888 -0.05897847
#>  [8,] -0.3570772  0.144244812 -0.005417102 -0.087514605  0.04597917 -0.06858307
#>  [9,] -0.2534600 -0.053390528 -0.097042115 -0.033880925 -0.05985811  0.01009129
#> [10,] -0.3367264  0.077415913 -0.024874162 -0.010436299  0.05293606 -0.01036568
#>                V35          V36           V37         V38         V39
#>  [1,]  0.090169429  0.061031754 -0.0152300316  0.08874064  0.06755874
#>  [2,] -0.102932218  0.016102720 -0.0228665226 -0.01413836 -0.10345450
#>  [3,] -0.149924374 -0.110798840 -0.1599730649 -0.05974423 -0.01937055
#>  [4,]  0.064626292  0.021682442 -0.0432747824 -0.02293689  0.07737424
#>  [5,]  0.021275606 -0.021166657  0.0563457861 -0.06069629  0.05320114
#>  [6,] -0.039621011 -0.027886132 -0.0915712881 -0.02422383  0.01092113
#>  [7,] -0.060771581  0.033837520 -0.0513102520 -0.02046084 -0.02367814
#>  [8,]  0.078657914 -0.055239065 -0.0880801969 -0.13948841 -0.13644648
#>  [9,] -0.002494227 -0.003871478  0.0260674628 -0.10294790 -0.06361033
#> [10,] -0.035169901 -0.129115163  0.0008143697 -0.15082679 -0.12218688
#>               V4           V40         V41         V42          V43         V44
#>  [1,] -0.3075639 -0.0425983148 -0.11751202  0.04352075 -0.135492167 -0.08642637
#>  [2,] -0.3663936 -0.0005415973  0.01542592 -0.08590427 -0.096239566 -0.16963845
#>  [3,] -0.2292708 -0.1135353015 -0.16059906 -0.06313387 -0.133269096 -0.09604485
#>  [4,] -0.1774767  0.0021402041 -0.07411157 -0.03968527 -0.042727601 -0.09791331
#>  [5,] -0.3064640 -0.0697072198 -0.04827771 -0.04312047 -0.086306300 -0.11901238
#>  [6,] -0.2877552 -0.1761653442 -0.16305441 -0.11915627 -0.042501265 -0.02851284
#>  [7,] -0.3177445 -0.0363227142 -0.04664561 -0.02888781  0.013616516 -0.11173921
#>  [8,] -0.2232724 -0.0485990994 -0.05084987 -0.06564533 -0.130983753 -0.17825153
#>  [9,] -0.2967960 -0.0138046192 -0.05741741 -0.19164782 -0.050259261 -0.20465902
#> [10,] -0.2406386 -0.1260994418 -0.04499531 -0.05806571  0.005761403 -0.08139817
#>               V45         V46         V47        V48        V49         V5
#>  [1,] -0.03102607 -0.06457911 -0.23416176 -0.1580833 -0.3272387 -0.1559800
#>  [2,] -0.08064235 -0.22102153 -0.23736377 -0.1787993 -0.3480053 -0.2173844
#>  [3,] -0.16266998 -0.16597120 -0.10769130 -0.2359204 -0.1576571 -0.2121488
#>  [4,] -0.08219804 -0.02773120 -0.07788955 -0.1796790 -0.1595256 -0.2328420
#>  [5,] -0.15524933 -0.03722213 -0.16427084 -0.1426564 -0.2840981 -0.1458347
#>  [6,] -0.21540488 -0.17187244 -0.18660550 -0.1644891 -0.3106635 -0.2822256
#>  [7,] -0.01573120 -0.06764412 -0.17682989 -0.3294301 -0.3364074 -0.1979436
#>  [8,] -0.05414512 -0.21377579 -0.22566906 -0.1616341 -0.2120818 -0.2194948
#>  [9,] -0.24630128 -0.26011852 -0.17379688 -0.2532727 -0.2481639 -0.2182066
#> [10,] -0.12539696 -0.14460469 -0.20078613 -0.2230996 -0.2142978 -0.2919363
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2599912 -0.2917095 -0.3129283 -0.2779018 -0.3729003 -0.2992797
#>  [2,] -0.3537943 -0.4030696 -0.2929925 -0.3318541 -0.3186148 -0.3899366
#>  [3,] -0.2498473 -0.2581220 -0.3498014 -0.3104837 -0.2808470 -0.3033696
#>  [4,] -0.3750933 -0.2387991 -0.4085847 -0.3319930 -0.3890143 -0.2711268
#>  [5,] -0.3381929 -0.3534651 -0.2726816 -0.2991944 -0.3446848 -0.2312036
#>  [6,] -0.2926604 -0.3337923 -0.4299378 -0.4150175 -0.3665823 -0.3699763
#>  [7,] -0.2859939 -0.2942633 -0.2919797 -0.3218908 -0.4021464 -0.4056999
#>  [8,] -0.3562907 -0.3937705 -0.2655507 -0.3594347 -0.3121816 -0.3795783
#>  [9,] -0.3482494 -0.1740713 -0.2322376 -0.3533329 -0.2419369 -0.3246721
#> [10,] -0.3900603 -0.3956588 -0.4235223 -0.3749067 -0.4221968 -0.4029304
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3857804 -0.2700440 -0.1930276 -0.3283697 -0.1437359 -0.3030272
#>  [2,] -0.3000763 -0.3542971 -0.4253976 -0.3926642 -0.1301457 -0.3604445
#>  [3,] -0.2408057 -0.3088610 -0.2184350 -0.3457179 -0.1817467 -0.2578097
#>  [4,] -0.2752059 -0.3057314 -0.3977207 -0.2303754 -0.1328465 -0.3565896
#>  [5,] -0.3301048 -0.3652025 -0.3280927 -0.4013556 -0.2758025 -0.2767123
#>  [6,] -0.2884213 -0.2888420 -0.3205080 -0.3664363 -0.2738970 -0.3124488
#>  [7,] -0.3464305 -0.3598648 -0.3601081 -0.4020222 -0.1925144 -0.2796510
#>  [8,] -0.4455531 -0.4194542 -0.3756379 -0.2878593 -0.3036928 -0.4020391
#>  [9,] -0.2552560 -0.2387613 -0.2939867 -0.2221327 -0.3026445 -0.2544613
#> [10,] -0.3899901 -0.4357894 -0.4243962 -0.3850687 -0.1075324 -0.4359459
#>                V7         V8          V9
#>  [1,] -0.09381000 -0.1415860 -0.14619012
#>  [2,] -0.30307421 -0.1435825 -0.05753523
#>  [3,] -0.27254994 -0.1761895 -0.16075692
#>  [4,] -0.16172589 -0.1493193 -0.14077513
#>  [5,] -0.17481802 -0.1102912 -0.08838076
#>  [6,] -0.13679917 -0.2521065 -0.14867138
#>  [7,] -0.19312406 -0.2426538 -0.21635874
#>  [8,] -0.09676307 -0.1723129 -0.20743783
#>  [9,] -0.28895366 -0.2692499 -0.18318704
#> [10,] -0.18629256 -0.1357460 -0.13542692
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]         [,4]        [,5]        [,6]
#> [1,] -0.1353503 -0.08083387 -0.07451169 -0.127034738 -0.02548694 -0.09470171
#> [2,]  0.1072487  0.08173231 -0.03526677 -0.006231506  0.09082880  0.14192377
#>             [,7]        [,8]        [,9]      [,10]
#> [1,] -0.06434602  0.01073263 -0.06733994 -0.1090458
#> [2,]  0.03191882 -0.01588249  0.14724085  0.1158730
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1          V10           V11           V12           V13
#>  [1,]  7.386103e-05 1.898403e-04  2.773283e-04  4.041995e-04  0.0004056174
#>  [2,]  2.487560e-05 1.523249e-04  1.560339e-04  2.198468e-04  0.0002588893
#>  [3,] -1.805266e-05 1.523117e-04  6.355811e-05  3.390547e-05  0.0001585157
#>  [4,]  5.061855e-06 1.282285e-04  1.067880e-04  1.459940e-04  0.0001906731
#>  [5,]  1.519523e-05 2.084745e-04  1.850398e-04  2.260046e-04  0.0003083005
#>  [6,]  1.364679e-04 3.132035e-04  5.024990e-04  6.963648e-04  0.0006650001
#>  [7,] -6.918698e-06 1.175034e-04  5.146822e-05  7.867747e-05  0.0001848359
#>  [8,] -1.347668e-04 1.241939e-05 -2.727988e-04 -4.029668e-04 -0.0001910893
#>  [9,]  6.530496e-05 3.566158e-05  1.451699e-04  2.862278e-04  0.0001983030
#> [10,]  1.089649e-04 2.784571e-04  4.130069e-04  5.618568e-04  0.0005632106
#>                 V14           V15           V16           V17           V18
#>  [1,]  2.984552e-04  0.0002626063  1.461763e-04 -1.995907e-05 -0.0001303598
#>  [2,]  3.221591e-04  0.0003494920  3.853661e-04  4.116150e-04  0.0004292234
#>  [3,]  4.903829e-04  0.0006669472  9.262615e-04  1.194402e-03  0.0013822046
#>  [4,]  3.105910e-04  0.0003633130  4.602105e-04  5.623018e-04  0.0006456242
#>  [5,]  4.961120e-04  0.0006029236  7.315371e-04  8.634884e-04  0.0009448318
#>  [6,]  3.164189e-04  0.0001652136 -1.897086e-04 -6.165616e-04 -0.0009344616
#>  [7,]  4.607921e-04  0.0006162699  7.974080e-04  9.834242e-04  0.0011051227
#>  [8,]  6.356460e-04  0.0010692951  1.757436e-03  2.514868e-03  0.0030142767
#>  [9,] -7.537063e-05 -0.0001642477 -4.827402e-04 -8.550617e-04 -0.0011329378
#> [10,]  3.438159e-04  0.0002522712  1.624384e-05 -2.802388e-04 -0.0005038545
#>                 V19            V2           V20           V21           V22
#>  [1,]  9.121848e-05  9.135155e-05  0.0003268266  0.0004858491  0.0005763432
#>  [2,]  4.232699e-04  4.068083e-05  0.0004213730  0.0004869649  0.0005462211
#>  [3,]  1.114861e-03  1.892449e-06  0.0008767710  0.0008460368  0.0009101854
#>  [4,]  5.634390e-04  1.985359e-05  0.0004740139  0.0004953287  0.0005331538
#>  [5,]  8.347005e-04  3.866711e-05  0.0007392200  0.0007951142  0.0008990416
#>  [6,] -3.906877e-04  1.554466e-04  0.0001475984  0.0004527970  0.0006257210
#>  [7,]  8.838486e-04  1.806356e-05  0.0006798814  0.0006983053  0.0008319352
#>  [8,]  2.093670e-03 -1.021331e-04  0.0011968482  0.0009603016  0.0010140020
#>  [9,] -8.368047e-04  7.415836e-05 -0.0005531811 -0.0003484849 -0.0001634637
#> [10,] -1.242640e-04  1.274243e-04  0.0002798443  0.0005090196  0.0006306570
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0004619614  0.0002850525  1.452490e-04  2.333112e-04  0.0005504403
#>  [2,]  0.0005719165  0.0005659886  5.063653e-04  5.935645e-04  0.0007002639
#>  [3,]  0.0010969701  0.0013059941  1.402303e-03  1.485269e-03  0.0012907089
#>  [4,]  0.0006118296  0.0006765494  6.543536e-04  7.324770e-04  0.0007450809
#>  [5,]  0.0009856194  0.0010528399  1.033892e-03  1.134972e-03  0.0011525851
#>  [6,]  0.0003376898 -0.0001112967 -4.242356e-04 -2.686030e-04  0.0004062273
#>  [7,]  0.0009790288  0.0010855055  1.088646e-03  1.190325e-03  0.0011548523
#>  [8,]  0.0015265361  0.0021390036  2.406851e-03  2.438354e-03  0.0017745437
#>  [9,] -0.0003856683 -0.0007850372 -1.117529e-03 -1.065378e-03 -0.0004966424
#> [10,]  0.0004285869  0.0001192797 -8.845188e-05  5.107116e-05  0.0005248930
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0006977155  0.0005274614  7.127138e-05  0.0004916893  0.0003367830
#>  [2,]  0.0006984579  0.0006228034  4.704597e-05  0.0005926706  0.0005782611
#>  [3,]  0.0010986863  0.0011549623  3.603965e-05  0.0011118820  0.0012058442
#>  [4,]  0.0006978199  0.0006739936  3.338615e-05  0.0006260948  0.0006464776
#>  [5,]  0.0010550288  0.0009922179  5.712056e-05  0.0009426006  0.0009501428
#>  [6,]  0.0006783168  0.0002875824  1.082658e-04  0.0002523258 -0.0001012539
#>  [7,]  0.0010138516  0.0009971366  4.254309e-05  0.0009675955  0.0010558306
#>  [8,]  0.0012546491  0.0015275931 -8.591712e-06  0.0014904287  0.0019682527
#>  [9,] -0.0003023150 -0.0006307309  3.559867e-05 -0.0005629054 -0.0007133754
#> [10,]  0.0007071238  0.0004189857  9.692408e-05  0.0004034093  0.0001548458
#>                 V32           V33           V34           V35           V36
#>  [1,]  2.959110e-04  8.234817e-05 -0.0000867088 -0.0003452186 -0.0005267799
#>  [2,]  5.249065e-04  5.102196e-04  0.0004681393  0.0004659938  0.0004790915
#>  [3,]  1.151591e-03  1.411447e-03  0.0015372823  0.0018541503  0.0020615791
#>  [4,]  5.861180e-04  6.557752e-04  0.0006752029  0.0007543129  0.0008471722
#>  [5,]  8.704969e-04  9.375607e-04  0.0009246627  0.0010292914  0.0011290179
#>  [6,] -2.180935e-04 -7.657930e-04 -0.0012001396 -0.0018248747 -0.0022043205
#>  [7,]  9.872151e-04  1.139769e-03  0.0011838618  0.0014033396  0.0015799980
#>  [8,]  1.936029e-03  2.665422e-03  0.0031116511  0.0040884449  0.0047806784
#>  [9,] -7.906873e-04 -1.245693e-03 -0.0016007536 -0.0020397480 -0.0022360814
#> [10,]  8.111853e-05 -2.911821e-04 -0.0005953498 -0.0010247220 -0.0013082472
#>                 V37           V38           V39            V4           V40
#>  [1,] -0.0003368849  2.106133e-05  8.665553e-05  7.625970e-05  4.290435e-05
#>  [2,]  0.0004688440  4.437623e-04  3.996302e-04  4.849456e-05  4.104804e-04
#>  [3,]  0.0017422611  1.138214e-03  8.448242e-04  3.449781e-05  9.287276e-04
#>  [4,]  0.0007761068  6.010051e-04  5.131770e-04  3.360885e-05  5.297455e-04
#>  [5,]  0.0010346234  7.889451e-04  6.394486e-04  5.525125e-05  6.986319e-04
#>  [6,] -0.0016484234 -6.627433e-04 -3.377529e-04  1.153088e-04 -4.146675e-04
#>  [7,]  0.0013928139  9.518909e-04  7.168344e-04  3.327416e-05  8.307818e-04
#>  [8,]  0.0040012327  2.355946e-03  1.666667e-03 -2.764081e-05  1.913922e-03
#>  [9,] -0.0016827088 -9.069546e-04 -6.414960e-04  2.968940e-05 -6.032853e-04
#> [10,] -0.0009578617 -3.012178e-04 -1.025179e-04  1.047928e-04 -1.585284e-04
#>                 V41          V42           V43           V44           V45
#>  [1,]  0.0003438089 0.0004817677  0.0004169805  3.609091e-04  5.395985e-04
#>  [2,]  0.0004490184 0.0004152116  0.0002861735  2.167013e-04  2.086827e-04
#>  [3,]  0.0006827521 0.0004128471  0.0001863718  9.895848e-05 -1.294003e-04
#>  [4,]  0.0004358335 0.0003443783  0.0002053132  1.431487e-04  5.917337e-05
#>  [5,]  0.0006433952 0.0005174606  0.0003145524  2.329644e-04  1.418876e-04
#>  [6,]  0.0002699393 0.0006534320  0.0006583127  6.048514e-04  1.053047e-03
#>  [7,]  0.0006948536 0.0004710248  0.0002222093  1.251363e-04 -2.932766e-05
#>  [8,]  0.0010174485 0.0002842442 -0.0001875822 -2.975604e-04 -9.904110e-04
#>  [9,] -0.0001407528 0.0001012364  0.0001385244  1.479431e-04  4.783200e-04
#> [10,]  0.0003643177 0.0006234383  0.0005819173  5.158109e-04  8.381905e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  5.358583e-04  3.406922e-04  2.093221e-04  1.243110e-04  9.291518e-05
#>  [2,]  1.809514e-04  1.356519e-04  1.056308e-04  5.403772e-05  6.689999e-05
#>  [3,] -1.939463e-04 -8.710396e-05 -5.265807e-06 -2.557865e-05  4.749923e-05
#>  [4,]  2.077518e-05  4.256736e-05  5.432675e-05  2.262982e-05  5.277736e-05
#>  [5,]  8.906128e-05  9.074702e-05  9.450879e-05  4.091405e-05  7.933151e-05
#>  [6,]  1.056787e-03  6.496165e-04  3.837590e-04  2.411367e-04  1.490905e-04
#>  [7,] -7.297126e-05 -1.047180e-05  3.578646e-05  6.072917e-07  4.347176e-05
#>  [8,] -1.079041e-03 -5.790264e-04 -2.578513e-04 -2.012664e-04 -2.597552e-05
#>  [9,]  5.543634e-04  3.441101e-04  1.841644e-04  1.245850e-04  2.714435e-05
#> [10,]  8.309867e-04  5.132681e-04  3.118682e-04  1.900088e-04  1.313032e-04
#>                 V50           V51           V52          V53           V54
#>  [1,]  3.621010e-05  4.221245e-05  3.346095e-05 1.502727e-05  9.955771e-06
#>  [2,]  2.368918e-05  2.272829e-05  1.483445e-05 1.365884e-05  1.299486e-05
#>  [3,]  1.953324e-05  3.458789e-06 -2.857364e-06 1.795511e-05  2.439304e-05
#>  [4,]  1.678206e-05  1.254159e-05  5.442150e-06 1.228606e-05  1.321700e-05
#>  [5,]  2.757907e-05  2.176187e-05  1.171009e-05 1.824534e-05  1.967008e-05
#>  [6,]  5.186499e-05  7.321349e-05  6.226138e-05 1.717753e-05  3.543527e-06
#>  [7,]  2.157481e-05  1.115368e-05  3.113080e-06 1.842807e-05  2.079269e-05
#>  [8,] -6.458659e-06 -4.181803e-05 -4.917196e-05 1.655974e-05  3.433608e-05
#>  [9,]  1.029106e-05  3.376857e-05  2.854581e-05 2.782391e-07 -1.277703e-05
#> [10,]  4.713191e-05  6.039981e-05  5.040997e-05 1.717439e-05  8.749020e-06
#>                 V55          V56           V57           V58           V59
#>  [1,]  8.075117e-06 1.312076e-05 -2.376178e-07  9.802553e-06  9.066715e-06
#>  [2,]  1.007970e-05 1.026696e-05  8.720155e-06  8.369553e-06  1.008596e-05
#>  [3,]  1.731590e-05 1.227583e-05  2.392191e-05  1.073039e-05  1.533857e-05
#>  [4,]  9.972122e-06 8.587255e-06  1.145656e-05  7.261790e-06  9.450314e-06
#>  [5,]  1.460665e-05 1.362967e-05  1.585337e-05  1.089628e-05  1.383256e-05
#>  [6,]  5.283377e-06 1.491823e-05 -1.471528e-05  1.145982e-05  6.109158e-06
#>  [7,]  1.497738e-05 1.262395e-05  1.834140e-05  9.212297e-06  1.338369e-05
#>  [8,]  2.227114e-05 9.804041e-06  4.811719e-05  7.377022e-06  1.973387e-05
#>  [9,] -8.440721e-06 3.764960e-07 -2.281388e-05 -4.307667e-06 -6.568077e-06
#> [10,]  8.430519e-06 1.503108e-05 -5.696296e-06  1.207834e-05  9.179664e-06
#>                  V6           V60            V7           V8            V9
#>  [1,]  9.386232e-05  9.973907e-06  7.413524e-05 1.232332e-04  2.239379e-04
#>  [2,]  9.236956e-05  1.004290e-05  8.872927e-05 9.961960e-05  1.329484e-04
#>  [3,]  1.211718e-04  1.561588e-05  1.478985e-04 1.008228e-04  7.501693e-05
#>  [4,]  8.724835e-05  9.029979e-06  9.165427e-05 8.520183e-05  8.621252e-05
#>  [5,]  1.289687e-04  1.430003e-05  1.355390e-04 1.328227e-04  1.587650e-04
#>  [6,]  1.065164e-04  6.801482e-06  5.882687e-05 2.000678e-04  4.198754e-04
#>  [7,]  1.047774e-04  1.501478e-05  1.103097e-04 8.422425e-05  6.465996e-05
#>  [8,]  1.234516e-04  2.101334e-05  1.807335e-04 1.226094e-05 -2.020712e-04
#>  [9,] -4.029706e-05 -2.606743e-06 -8.461663e-05 3.090785e-05  1.293694e-04
#> [10,]  1.122010e-04  9.588001e-06  7.828573e-05 1.784023e-04  3.490017e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]      [,5]        [,6]
#> [1,]  0.07415928  0.07416496  0.04131653  0.07395752  0.062539  0.05798003
#> [2,] -0.07415928 -0.07416496 -0.04131653 -0.07395752 -0.062539 -0.05798003
#>             [,7]       [,8]        [,9]       [,10]
#> [1,]  0.05549717  0.0628982  0.04765111  0.06296599
#> [2,] -0.05549717 -0.0628982 -0.04765111 -0.06296599
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5497213 0.7449581 0.3263149 0.6989653 0.4851753 0.5521357 0.3763153
#>  [8] 0.6401463 0.2532762 0.6705259
#> 
#> $B[[2]]
#> [1] -0.06433156  0.07660248
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0004132038  0.0010615111  0.0024092213  0.0012886518  0.0018565093
#>  [6] -0.0005175959  0.0020367032  0.0042321993 -0.0015998099  0.0000367115
#> 
#> $vB[[2]]
#> [1]  0.08420664 -0.08420664
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#>  [2,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [3,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#>  [4,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#>  [5,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#>  [6,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#>  [7,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [8,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#>  [9,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [10,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [11,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [12,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [13,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [14,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [15,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [16,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [17,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [18,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [19,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [20,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [21,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [22,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [23,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [24,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [25,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [26,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [27,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [28,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [29,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [30,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [31,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [32,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [33,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [34,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [35,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [36,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [37,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [38,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [39,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [40,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#>  [2,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [3,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#>  [4,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#>  [5,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#>  [6,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#>  [7,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [8,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#>  [9,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [10,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [11,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [12,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [13,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [14,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [15,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [16,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [17,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [18,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [19,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [20,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [21,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [22,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [23,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [24,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [25,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [26,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [27,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [28,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [29,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [30,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [31,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [32,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [33,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [34,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [35,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [36,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [37,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [38,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [39,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [40,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#>  [2,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [3,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#>  [4,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#>  [5,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#>  [6,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#>  [7,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [8,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#>  [9,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [10,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [11,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [12,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [13,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [14,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [15,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [16,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [17,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [18,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [19,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [20,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [21,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [22,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [23,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [24,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [25,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [26,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [27,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [28,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [29,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [30,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [31,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [32,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [33,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [34,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [35,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [36,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [37,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [38,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [39,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [40,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#>  [2,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [3,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#>  [4,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#>  [5,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#>  [6,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#>  [7,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [8,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#>  [9,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [10,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [11,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [12,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [13,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [14,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [15,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [16,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [17,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [18,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [19,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [20,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [21,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [22,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [23,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [24,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [25,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [26,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [27,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [28,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [29,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [30,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [31,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [32,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [33,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [34,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [35,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [36,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [37,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [38,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [39,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [40,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#>  [2,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [3,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#>  [4,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#>  [5,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#>  [6,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#>  [7,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [8,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#>  [9,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [10,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [11,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [12,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [13,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [14,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [15,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [16,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [17,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [18,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [19,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [20,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [21,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [22,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [23,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [24,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [25,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [26,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [27,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [28,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [29,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [30,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [31,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [32,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [33,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [34,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [35,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [36,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [37,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [38,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [39,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [40,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#>  [2,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [3,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#>  [4,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#>  [5,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#>  [6,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#>  [7,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [8,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#>  [9,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [10,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [11,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [12,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [13,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [14,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [15,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [16,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [17,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [18,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [19,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [20,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [21,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [22,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [23,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [24,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [25,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [26,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [27,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [28,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [29,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [30,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [31,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [32,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [33,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [34,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [35,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [36,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [37,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [38,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [39,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [40,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9169048 0.8960888 0.5032963 0.9270090 0.7665937 0.6860961 0.6404101
#>  [2,] 0.9047164 0.8997526 0.5199364 0.9271946 0.7720409 0.7056679 0.6397797
#>  [3,] 0.8680183 0.8597092 0.4890580 0.9021612 0.7150464 0.6726731 0.6136688
#>  [4,] 0.8592465 0.8734388 0.5528975 0.9061518 0.7464841 0.7049834 0.6176588
#>  [5,] 0.8581498 0.9071543 0.5125564 0.9147234 0.7421521 0.6440028 0.5591161
#>  [6,] 0.8976615 0.9386884 0.6050685 0.9474526 0.7977728 0.7712730 0.6605946
#>  [7,] 0.8237709 0.8084303 0.3831768 0.8826777 0.6524454 0.5398385 0.5140268
#>  [8,] 0.8085467 0.8650106 0.5176789 0.8930287 0.7106438 0.6453989 0.5711464
#>  [9,] 0.8443805 0.8098966 0.3424339 0.8742880 0.6533264 0.5054020 0.4947982
#> [10,] 0.8567088 0.8315170 0.4213617 0.8796922 0.7127402 0.6125254 0.5408290
#> [11,] 0.8735003 0.8930675 0.4105875 0.9099699 0.7187368 0.6026313 0.5417349
#> [12,] 0.8950133 0.9270419 0.5667797 0.9456520 0.7745160 0.7312167 0.6321702
#> [13,] 0.8291051 0.8519771 0.5601079 0.8854698 0.7102773 0.7142021 0.6375182
#> [14,] 0.8596198 0.9016060 0.4920529 0.9202686 0.7245653 0.6757909 0.5763801
#> [15,] 0.8611250 0.8006021 0.3444280 0.8717727 0.6656820 0.5468659 0.5219838
#> [16,] 0.8327630 0.8007729 0.4719083 0.8328429 0.7148802 0.6035816 0.5666769
#> [17,] 0.8244863 0.8560729 0.5789065 0.8864216 0.7139377 0.7205597 0.6523313
#> [18,] 0.9190716 0.8934448 0.3992957 0.9273228 0.7502033 0.6447797 0.5778639
#> [19,] 0.7708889 0.7491046 0.3311518 0.8076487 0.6071154 0.4982629 0.4838219
#> [20,] 0.8054747 0.7853114 0.3748204 0.8384221 0.6296364 0.5293056 0.5040577
#> [21,] 0.8544796 0.8658349 0.4885368 0.8959016 0.7162336 0.6481481 0.5903121
#> [22,] 0.8983938 0.8929951 0.5298761 0.9171840 0.7799631 0.6899737 0.6312804
#> [23,] 0.8639579 0.8794384 0.4791066 0.9051226 0.7216088 0.6173876 0.5653347
#> [24,] 0.8825618 0.8923699 0.4779641 0.9182236 0.7230141 0.6572392 0.5967706
#> [25,] 0.8349545 0.8349620 0.4681348 0.8731422 0.7098968 0.6273026 0.5603180
#> [26,] 0.8893882 0.8747903 0.5022910 0.9040501 0.7517511 0.6872999 0.6272278
#> [27,] 0.7647664 0.8548911 0.4129464 0.8584726 0.6413767 0.5238422 0.4405295
#> [28,] 0.8154248 0.8792199 0.4257049 0.9025102 0.6795575 0.6113330 0.5693411
#> [29,] 0.8257312 0.8952726 0.5595292 0.9007339 0.7311424 0.6772728 0.6022062
#> [30,] 0.8819147 0.8960667 0.5723927 0.9228047 0.7623794 0.7315571 0.6497619
#> [31,] 0.8060716 0.8714946 0.4378791 0.8839555 0.6644201 0.6094696 0.5538792
#> [32,] 0.8408615 0.8819331 0.4793945 0.8939483 0.7141871 0.6101338 0.5591940
#> [33,] 0.8892497 0.9142884 0.5939999 0.9358037 0.7707258 0.7419393 0.6649452
#> [34,] 0.8483399 0.8363608 0.3521123 0.8674170 0.6687663 0.5432574 0.5106811
#> [35,] 0.8411350 0.8729011 0.3733185 0.8986994 0.6591005 0.5971986 0.5167575
#> [36,] 0.8409666 0.8509232 0.4882914 0.8794949 0.7175282 0.6475442 0.5698973
#> [37,] 0.8896442 0.9158375 0.5971231 0.9394421 0.7838246 0.7491861 0.6495848
#> [38,] 0.8708514 0.9115046 0.5305695 0.9343316 0.7448626 0.6807772 0.6150181
#> [39,] 0.8747714 0.8434091 0.4620770 0.8798782 0.7224702 0.6281923 0.5939668
#> [40,] 0.8620623 0.8823937 0.4799197 0.9136354 0.7364485 0.6357064 0.5539723
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7734879 0.4257868 0.8033342
#>  [2,] 0.7541251 0.4376137 0.7919250
#>  [3,] 0.7187161 0.4246123 0.7630817
#>  [4,] 0.7265571 0.3967862 0.7884344
#>  [5,] 0.7046821 0.3074425 0.7762697
#>  [6,] 0.7813899 0.4106885 0.8505195
#>  [7,] 0.5918817 0.2650549 0.6587117
#>  [8,] 0.6886039 0.3708021 0.7452161
#>  [9,] 0.5903826 0.2106486 0.6608039
#> [10,] 0.6469169 0.3509566 0.7013397
#> [11,] 0.6608311 0.3154280 0.7548283
#> [12,] 0.7587215 0.3454064 0.8276656
#> [13,] 0.7397410 0.4628815 0.7902876
#> [14,] 0.6675820 0.2919980 0.7813995
#> [15,] 0.5923493 0.2697769 0.6636525
#> [16,] 0.6571514 0.4318739 0.6962914
#> [17,] 0.7452221 0.4844129 0.7912034
#> [18,] 0.6804586 0.3136434 0.7410565
#> [19,] 0.5256191 0.2809641 0.5993935
#> [20,] 0.5854979 0.2596043 0.6569270
#> [21,] 0.6936951 0.3569766 0.7554140
#> [22,] 0.7550610 0.4336732 0.7837940
#> [23,] 0.6723689 0.3292267 0.7138189
#> [24,] 0.7239499 0.3581409 0.7811243
#> [25,] 0.6722548 0.3333439 0.7479024
#> [26,] 0.7386714 0.4264031 0.7745097
#> [27,] 0.5437892 0.2035339 0.6697883
#> [28,] 0.6270138 0.2878942 0.7234453
#> [29,] 0.7150734 0.4031072 0.7659897
#> [30,] 0.7482774 0.4159700 0.8084573
#> [31,] 0.6292195 0.2904546 0.7113010
#> [32,] 0.6669100 0.3381706 0.7279260
#> [33,] 0.7753287 0.4208548 0.8240374
#> [34,] 0.5961302 0.2953883 0.6762543
#> [35,] 0.5567154 0.2332557 0.6953536
#> [36,] 0.6707418 0.3900998 0.7390366
#> [37,] 0.7752794 0.4026891 0.8330182
#> [38,] 0.7248794 0.3498452 0.7875238
#> [39,] 0.7042764 0.3836364 0.7435340
#> [40,] 0.6636750 0.2758682 0.7521687
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4740707 0.5259293
#>  [2,] 0.4731404 0.5268596
#>  [3,] 0.4741671 0.5258329
#>  [4,] 0.4746532 0.5253468
#>  [5,] 0.4779184 0.5220816
#>  [6,] 0.4727756 0.5272244
#>  [7,] 0.4801105 0.5198895
#>  [8,] 0.4773469 0.5226531
#>  [9,] 0.4816019 0.5183981
#> [10,] 0.4761721 0.5238279
#> [11,] 0.4762264 0.5237736
#> [12,] 0.4751468 0.5248532
#> [13,] 0.4735822 0.5264178
#> [14,] 0.4757096 0.5242904
#> [15,] 0.4783677 0.5216323
#> [16,] 0.4755498 0.5244502
#> [17,] 0.4733025 0.5266975
#> [18,] 0.4752208 0.5247792
#> [19,] 0.4800078 0.5199922
#> [20,] 0.4804274 0.5195726
#> [21,] 0.4763207 0.5236793
#> [22,] 0.4741645 0.5258355
#> [23,] 0.4778977 0.5221023
#> [24,] 0.4758482 0.5241518
#> [25,] 0.4770917 0.5229083
#> [26,] 0.4738801 0.5261199
#> [27,] 0.4814084 0.5185916
#> [28,] 0.4778170 0.5221830
#> [29,] 0.4760720 0.5239280
#> [30,] 0.4734773 0.5265227
#> [31,] 0.4783446 0.5216554
#> [32,] 0.4777171 0.5222829
#> [33,] 0.4738216 0.5261784
#> [34,] 0.4777393 0.5222607
#> [35,] 0.4764761 0.5235239
#> [36,] 0.4750702 0.5249298
#> [37,] 0.4739361 0.5260639
#> [38,] 0.4762644 0.5237356
#> [39,] 0.4760585 0.5239415
#> [40,] 0.4777476 0.5222524
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]     [,2]         [,3]     [,4]      [,5]         [,6]
#>  [1,] 2.401017 2.154503  0.013185483 2.541628 1.1891763  0.781930628
#>  [2,] 2.250764 2.194478  0.079787735 2.544373 1.2198711  0.874435834
#>  [3,] 1.883550 1.812876 -0.043775130 2.221472 0.9200210  0.720300331
#>  [4,] 1.809045 1.931712  0.212384684 2.267529 1.0799480  0.871142662
#>  [5,] 1.800007 2.279374  0.050235966 2.372722 1.0571844  0.592780406
#>  [6,] 2.171507 2.728514  0.426629546 2.892061 1.3724320  1.215513220
#>  [7,] 1.542107 1.439843 -0.476086047 2.018036 0.6298058  0.159692441
#>  [8,] 1.440595 1.857546  0.070745015 2.122058 0.8985127  0.598875126
#>  [9,] 1.691189 1.449338 -0.652466695 1.939416 0.6336930  0.021608884
#> [10,] 1.788220 1.596417 -0.317186059 1.989519 0.9087299  0.457940252
#> [11,] 1.932268 2.122464 -0.361537361 2.313267 0.9382045  0.416440881
#> [12,] 2.143005 2.542113  0.268724169 2.856466 1.2339890  1.000804215
#> [13,] 1.579298 1.750193  0.241600040 2.045280 0.8967310  0.915880908
#> [14,] 1.812136 2.215198 -0.031791144 2.446002 0.9672214  0.734495012
#> [15,] 1.824665 1.390062 -0.643622860 1.916724 0.6887193  0.188015468
#> [16,] 1.605337 1.391132 -0.112485414 1.605911 0.9192054  0.420410817
#> [17,] 1.547044 1.783049  0.318286206 2.054699 0.9145860  0.947239817
#> [18,] 2.429799 2.126422 -0.408400755 2.546275 1.0996968  0.596170541
#> [19,] 1.213337 1.093842 -0.702980236 1.434804 0.4352028 -0.006948478
#> [20,] 1.420869 1.296891 -0.511591853 1.646534 0.5306574  0.117356801
#> [21,] 1.770176 1.864623 -0.045860636 2.152494 0.9258550  0.610908708
#> [22,] 2.179504 2.121706  0.119647122 2.404686 1.2654512  0.799996421
#> [23,] 1.848560 1.987123 -0.083622457 2.255485 0.9524559  0.478474615
#> [24,] 2.016916 2.115181 -0.088200932 2.418452 0.9594621  0.651015349
#> [25,] 1.621156 1.621210 -0.127633630 1.929032 0.8948827  0.520662025
#> [26,] 2.084507 1.943994  0.009163948 2.243059 1.1079732  0.787526205
#> [27,] 1.178992 1.773490 -0.351798160 1.802661 0.5813443  0.095441041
#> [28,] 1.485652 1.985063 -0.299397128 2.225432 0.7517390  0.452918858
#> [29,] 1.555671 2.145767  0.239251664 2.205406 1.0004263  0.741266735
#> [30,] 2.010688 2.154266  0.291619949 2.481080 1.1657692  1.002537162
#> [31,] 1.424683 1.914239 -0.249773989 2.030433 0.6830544  0.445083397
#> [32,] 1.664652 2.010865 -0.082468745 2.131722 0.9158075  0.447874777
#> [33,] 2.083099 2.367158  0.380525837 2.679460 1.2124141  1.056072600
#> [34,] 1.721640 1.631396 -0.609767269 1.878311 0.7026104  0.173463358
#> [35,] 1.666698 1.926857 -0.518006292 2.182856 0.6592881  0.393806226
#> [36,] 1.665437 1.741860 -0.046842813 1.987656 0.9322335  0.608261884
#> [37,] 2.087112 2.387089  0.393492225 2.741687 1.2880953  1.094276234
#> [38,] 1.908507 2.332146  0.122430676 2.655213 1.0713977  0.757345759
#> [39,] 1.943822 1.683816 -0.151983839 1.991277 0.9567478  0.524469479
#> [40,] 1.832525 2.015296 -0.080364496 2.358853 1.0275903  0.556776628
#>              [,7]      [,8]        [,9]     [,10]
#>  [1,]  0.57714436 1.2281119 -0.29906212 1.4072647
#>  [2,]  0.57440822 1.1207352 -0.25085246 1.3365684
#>  [3,]  0.46276027 0.9381019 -0.30386748 1.1696497
#>  [4,]  0.47962269 0.9772241 -0.41887395 1.3155140
#>  [5,]  0.23757562 0.8696944 -0.81210316 1.2440586
#>  [6,]  0.66594514 1.2737844 -0.36111981 1.7386814
#>  [7,]  0.05612199 0.3717499 -1.01985903 0.6575582
#>  [8,]  0.28653010 0.7936004 -0.52877740 1.0732585
#>  [9,] -0.02080797 0.3655472 -1.32102021 0.6668789
#> [10,]  0.16368051 0.6055144 -0.61483704 0.8536856
#> [11,]  0.16732887 0.6670002 -0.77486354 1.1245318
#> [12,]  0.54153853 1.1456827 -0.63929277 1.5691727
#> [13,]  0.56460866 1.0446230 -0.14874756 1.3266598
#> [14,]  0.30793078 0.6972691 -0.88569978 1.2738404
#> [15,]  0.08799185 0.3736859 -0.99575482 0.6796138
#> [16,]  0.26830577 0.6506254 -0.27420979 0.8296998
#> [17,]  0.62930237 1.0732900 -0.06236854 1.3321947
#> [18,]  0.31401060 0.7558800 -0.78314046 1.0514671
#> [19,] -0.06473505 0.1025661 -0.93968424 0.4029385
#> [20,]  0.01623105 0.3453844 -1.04802606 0.6496297
#> [21,]  0.36525583 0.8174515 -0.58851074 1.1276990
#> [22,]  0.53771361 1.1257894 -0.26688018 1.2879148
#> [23,]  0.26284178 0.7189191 -0.71168452 0.9140045
#> [24,]  0.39202719 0.9641400 -0.58344223 1.2722304
#> [25,]  0.24245289 0.7184010 -0.69309943 1.0874563
#> [26,]  0.52034242 1.0390743 -0.29654190 1.2339533
#> [27,] -0.23901341 0.1756067 -1.36435227 0.7072278
#> [28,]  0.27916352 0.5194270 -0.90563359 0.9616164
#> [29,]  0.41466616 0.9201538 -0.39253509 1.1858035
#> [30,]  0.61799289 1.0894462 -0.33933925 1.4400173
#> [31,]  0.21635675 0.5288700 -0.89317728 0.9017110
#> [32,]  0.23789164 0.6942424 -0.67145751 0.9841251
#> [33,]  0.68541048 1.2386484 -0.31926532 1.5439446
#> [34,]  0.04273098 0.3893666 -0.86935587 0.7366110
#> [35,]  0.06705505 0.2278420 -1.19001825 0.8252689
#> [36,]  0.28143227 0.7115420 -0.44689276 1.0409671
#> [37,]  0.61721467 1.2383654 -0.39427311 1.6071703
#> [38,]  0.46845554 0.9687958 -0.61971967 1.3100638
#> [39,]  0.38038855 0.8677456 -0.47414183 1.0644182
#> [40,]  0.21673344 0.6797146 -0.96505023 1.1102125
#> 
#> $pre[[3]]
#>              [,1]          [,2]
#>  [1,] -0.09902301  0.0047871871
#>  [2,] -0.10050798  0.0070339103
#>  [3,] -0.09733283  0.0060910514
#>  [4,] -0.09931861  0.0021556083
#>  [5,] -0.09528640 -0.0069025765
#>  [6,] -0.10603942  0.0029658840
#>  [7,] -0.08768237 -0.0080823292
#>  [8,] -0.09331419 -0.0026396152
#>  [9,] -0.08482374 -0.0111982128
#> [10,] -0.09123655  0.0041472894
#> [11,] -0.09408713  0.0010790658
#> [12,] -0.10290305 -0.0034081254
#> [13,] -0.09874450  0.0070252365
#> [14,] -0.09958876 -0.0023504718
#> [15,] -0.08817507 -0.0015919832
#> [16,] -0.08918549  0.0086935300
#> [17,] -0.09945520  0.0074363531
#> [18,] -0.09601105  0.0031869042
#> [19,] -0.08068593 -0.0006742885
#> [20,] -0.08443289 -0.0061026035
#> [21,] -0.09520175 -0.0004134722
#> [22,] -0.09840841  0.0050258140
#> [23,] -0.09360628 -0.0051394654
#> [24,] -0.09705063 -0.0003683303
#> [25,] -0.09188698 -0.0001894828
#> [26,] -0.09767023  0.0069047638
#> [27,] -0.08609771 -0.0116970811
#> [28,] -0.09306424 -0.0042741008
#> [29,] -0.09669584 -0.0009106855
#> [30,] -0.10275967  0.0034306794
#> [31,] -0.09154370 -0.0048679056
#> [32,] -0.09280846 -0.0036177790
#> [33,] -0.10376588  0.0010434494
#> [34,] -0.08811485  0.0009867086
#> [35,] -0.09523363 -0.0010685847
#> [36,] -0.09473322  0.0050687823
#> [37,] -0.10380180  0.0005482438
#> [38,] -0.09934228 -0.0043285112
#> [39,] -0.09283186  0.0030075551
#> [40,] -0.09536786 -0.0062993362
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4740707  0.4740707
#>  [2,] -0.4731404  0.4731404
#>  [3,] -0.4741671  0.4741671
#>  [4,]  0.5253468 -0.5253468
#>  [5,]  0.5220816 -0.5220816
#>  [6,] -0.4727756  0.4727756
#>  [7,]  0.5198895 -0.5198895
#>  [8,]  0.5226531 -0.5226531
#>  [9,]  0.5183981 -0.5183981
#> [10,] -0.4761721  0.4761721
#> [11,] -0.4762264  0.4762264
#> [12,]  0.5248532 -0.5248532
#> [13,] -0.4735822  0.4735822
#> [14,]  0.5242904 -0.5242904
#> [15,]  0.5216323 -0.5216323
#> [16,] -0.4755498  0.4755498
#> [17,] -0.4733025  0.4733025
#> [18,] -0.4752208  0.4752208
#> [19,]  0.5199922 -0.5199922
#> [20,]  0.5195726 -0.5195726
#> [21,]  0.5236793 -0.5236793
#> [22,] -0.4741645  0.4741645
#> [23,] -0.4778977  0.4778977
#> [24,] -0.4758482  0.4758482
#> [25,]  0.5229083 -0.5229083
#> [26,] -0.4738801  0.4738801
#> [27,] -0.4814084  0.4814084
#> [28,] -0.4778170  0.4778170
#> [29,] -0.4760720  0.4760720
#> [30,] -0.4734773  0.4734773
#> [31,] -0.4783446  0.4783446
#> [32,] -0.4777171  0.4777171
#> [33,]  0.5261784 -0.5261784
#> [34,] -0.4777393  0.4777393
#> [35,] -0.4764761  0.4764761
#> [36,] -0.4750702  0.4750702
#> [37,]  0.5260639 -0.5260639
#> [38,]  0.5237356 -0.5237356
#> [39,] -0.4760585  0.4760585
#> [40,]  0.5222524 -0.5222524
#> 
#> $L
#> [1] 0.7087825 0.8251758 0.9088513 0.7147096 0.6926374 0.6832555
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174