Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13           V14
#>  [1,] -0.2655766 -0.11433272 -0.09302365 -0.15004134 -0.159261403 -5.460731e-02
#>  [2,] -0.3928976 -0.18425150 -0.03231045 -0.11463214 -0.017528035 -4.102970e-02
#>  [3,] -0.2090596 -0.08061917 -0.13034010  0.03514359 -0.032568290 -9.931679e-05
#>  [4,] -0.2965592 -0.17356913 -0.07882769 -0.15520577  0.008079134 -2.854244e-02
#>  [5,] -0.4148850 -0.02140899 -0.15957919 -0.08375839 -0.037691833 -1.598904e-01
#>  [6,] -0.3417555 -0.06877683 -0.12867706 -0.07428794  0.076509797  4.885703e-02
#>  [7,] -0.2999109 -0.19623845 -0.04377878 -0.15884638 -0.039875246 -9.751746e-02
#>  [8,] -0.1744050 -0.09648439 -0.08008696 -0.09659456 -0.161329597 -1.225400e-01
#>  [9,] -0.3555303 -0.03825749 -0.10892200 -0.06564621 -0.086461095 -8.191441e-02
#> [10,] -0.3407294 -0.15929987 -0.12190545 -0.14538559 -0.124908345 -1.280520e-01
#>               V15         V16         V17          V18         V19         V2
#>  [1,]  0.01829559 -0.14650257  0.04762935 -0.027341324  0.07082201 -0.2191021
#>  [2,] -0.09780910  0.01818350  0.09434211 -0.002233646  0.18841336 -0.2404236
#>  [3,] -0.07926555 -0.03809605  0.12252335  0.120978938  0.03487702 -0.3017882
#>  [4,]  0.05776908  0.04082469 -0.01967147 -0.086071963 -0.04085875 -0.2131619
#>  [5,] -0.02263408 -0.10289814 -0.06816879  0.042583381 -0.08525927 -0.4112101
#>  [6,] -0.02778394 -0.02448490  0.04656106  0.033629512  0.01395791 -0.2834437
#>  [7,] -0.04010343 -0.20391899 -0.14250203 -0.025626470 -0.05159185 -0.1972548
#>  [8,] -0.16916998 -0.08186549 -0.16788077 -0.117226049 -0.08227754 -0.2675012
#>  [9,] -0.01005883  0.07008706  0.08845198  0.183538826  0.05193609 -0.3729229
#> [10,] -0.12321259  0.02218245 -0.10406288 -0.025433351  0.13763733 -0.2698081
#>                V20         V21          V22         V23         V24         V25
#>  [1,] -0.080756069  0.06911892 -0.024823201  0.13572110  0.08758772  0.11108104
#>  [2,]  0.219171640  0.10971062  0.074889039  0.03754163  0.05102223  0.12595491
#>  [3,] -0.009196265  0.01319653  0.067537603  0.10590559  0.18527236  0.03445931
#>  [4,] -0.013711101  0.06134903  0.011516078 -0.03742433  0.14909550  0.07655087
#>  [5,] -0.045907622  0.12523059  0.152780156  0.17758353  0.19258708  0.07851246
#>  [6,]  0.105190882  0.08319475  0.115607407  0.24431217  0.29400093  0.24689959
#>  [7,]  0.033582695  0.09391344  0.001683061  0.03931451  0.05618111 -0.08831996
#>  [8,] -0.054549330 -0.02131078  0.078634593  0.04475965 -0.03383516 -0.03884991
#>  [9,]  0.140945853  0.19363402  0.139927894  0.11806536  0.09026976  0.19814763
#> [10,]  0.032018815  0.06458791  0.217333546  0.17581963  0.13350375  0.20160838
#>                V26         V27          V28         V29         V3          V30
#>  [1,]  0.092434784  0.11362152  0.006265500  0.13462137 -0.2360676  0.003317546
#>  [2,]  0.042113951  0.10541264  0.067705838  0.19876038 -0.3434698  0.088772365
#>  [3,]  0.131923359  0.06147863  0.147565746  0.06557762 -0.2553988  0.102589113
#>  [4,] -0.002183884  0.10923736  0.159659631 -0.03925368 -0.1877937 -0.120751031
#>  [5,]  0.138733607  0.10060784  0.014371816 -0.03057698 -0.3112733  0.034812133
#>  [6,]  0.304162014  0.27065787  0.250791247  0.16097062 -0.4187703  0.233739922
#>  [7,]  0.101611045 -0.05833107 -0.003298952 -0.06152663 -0.1834935 -0.032258136
#>  [8,]  0.143701384  0.04895312  0.151397767  0.15109237 -0.1798151  0.154207311
#>  [9,]  0.228209639  0.15721279  0.108351726  0.08472180 -0.1904007  0.059911038
#> [10,]  0.048666744  0.13218870  0.105705553  0.14982856 -0.2962949  0.080142816
#>                V31         V32           V33          V34         V35
#>  [1,]  0.050979436 -0.06981597 -0.0544125362 -0.120636935 -0.18142942
#>  [2,] -0.008580219  0.12511016 -0.0298199391  0.147891627  0.12368701
#>  [3,]  0.024654469 -0.01384522 -0.0047478138 -0.129848962 -0.04224605
#>  [4,]  0.024451573 -0.04847770 -0.1281485239 -0.155661823 -0.22330658
#>  [5,] -0.043248083 -0.14163944 -0.0491843462  0.003803346 -0.09224948
#>  [6,]  0.203444593  0.11422628  0.1456063683  0.064553414  0.08557379
#>  [7,] -0.124229478 -0.03591820 -0.0006449336 -0.077013153 -0.16545975
#>  [8,] -0.002383258  0.05190037 -0.1032171337  0.078072202  0.04559793
#>  [9,] -0.014514084 -0.03769140  0.0855306565  0.063135361 -0.03917986
#> [10,]  0.138208696 -0.10427726  0.0140597296  0.082205010  0.04176220
#>               V36         V37          V38         V39         V4          V40
#>  [1,] -0.15796938 -0.03649037 -0.056888075 -0.15581659 -0.3152094 -0.072062656
#>  [2,]  0.07390510  0.07880291  0.017583784  0.01520310 -0.3061476 -0.008651298
#>  [3,] -0.15165212 -0.03360845 -0.118229821 -0.06276955 -0.1670820  0.010969095
#>  [4,] -0.19626294 -0.20590456 -0.077879511 -0.13610279 -0.2460972 -0.136525592
#>  [5,] -0.06323058 -0.17325157 -0.160370811 -0.02250276 -0.3740498 -0.156484374
#>  [6,] -0.08982011 -0.04609993 -0.152304630 -0.09982076 -0.2093656  0.045439781
#>  [7,] -0.05054482 -0.19047236 -0.186454464 -0.16478123 -0.2720333 -0.110777038
#>  [8,]  0.01982145 -0.13857994 -0.079207231 -0.13337852 -0.1579210 -0.008155588
#>  [9,]  0.16648405  0.09591360  0.074772205  0.01302716 -0.2135225  0.015666994
#> [10,]  0.03813236 -0.07488559 -0.003174845  0.04810614 -0.2738944 -0.106691547
#>               V41          V42         V43         V44         V45         V46
#>  [1,] -0.21756038 -0.125753839 -0.16704396 -0.10549132 -0.15239241 -0.10220090
#>  [2,] -0.04122640 -0.033837147 -0.02432047 -0.17730628 -0.15610806 -0.07625421
#>  [3,] -0.04559507 -0.077349314 -0.05190411 -0.16722416 -0.23936466 -0.18804725
#>  [4,] -0.03592547 -0.014986349 -0.20973660 -0.09787359 -0.15701304 -0.24514345
#>  [5,] -0.09959285 -0.087953268 -0.17058770 -0.20966634 -0.18721822 -0.17158412
#>  [6,] -0.01658567  0.007413771 -0.08980536 -0.13376619 -0.15993310 -0.14986980
#>  [7,] -0.18430332 -0.150619448 -0.09451656 -0.15652336 -0.22011701 -0.09217721
#>  [8,] -0.13448880 -0.106868943 -0.07936188 -0.14688681 -0.16420109 -0.22202801
#>  [9,]  0.08062389 -0.030622619 -0.10080940 -0.02178914 -0.07401705 -0.12617895
#> [10,] -0.09382891 -0.139406704  0.01385377 -0.06804697 -0.09890882 -0.12356539
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.2808239 -0.1546987 -0.3632602 -0.2577639 -0.3384505 -0.3930533
#>  [2,] -0.1781560 -0.1699815 -0.2142715 -0.1925750 -0.4100991 -0.4069809
#>  [3,] -0.1469211 -0.1734486 -0.2360738 -0.2271298 -0.2229828 -0.3360835
#>  [4,] -0.2074844 -0.2753074 -0.3202389 -0.2578886 -0.3451971 -0.3039538
#>  [5,] -0.2189194 -0.2411349 -0.2431857 -0.2412735 -0.3557650 -0.2545769
#>  [6,] -0.2119237 -0.1863509 -0.2640639 -0.3396611 -0.3338870 -0.3259768
#>  [7,] -0.1649801 -0.1740462 -0.2211877 -0.2905124 -0.1941414 -0.2904385
#>  [8,] -0.2286757 -0.2616168 -0.1904395 -0.2038162 -0.2537879 -0.3282797
#>  [9,] -0.1778536 -0.1817425 -0.3577447 -0.1894059 -0.3442649 -0.2564662
#> [10,] -0.1099723 -0.3258667 -0.2458336 -0.2049964 -0.3798195 -0.4276230
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3556025 -0.3060934 -0.3104609 -0.3536479 -0.2754971 -0.2864167
#>  [2,] -0.2981623 -0.3037885 -0.4153530 -0.4157596 -0.3091183 -0.4111974
#>  [3,] -0.3381869 -0.1784414 -0.2375555 -0.3782165 -0.2691931 -0.3436912
#>  [4,] -0.3298390 -0.3241933 -0.3350999 -0.3969743 -0.2889053 -0.3585437
#>  [5,] -0.3229133 -0.3469280 -0.4050943 -0.3306161 -0.4025584 -0.3184537
#>  [6,] -0.3896294 -0.4699002 -0.4368619 -0.3811189 -0.4957506 -0.3667952
#>  [7,] -0.3406767 -0.3498703 -0.3560188 -0.2456108 -0.1765726 -0.2152491
#>  [8,] -0.2714123 -0.2553526 -0.1997569 -0.2664725 -0.3854539 -0.2442512
#>  [9,] -0.2956812 -0.3567403 -0.3980033 -0.3429071 -0.4398974 -0.4319586
#> [10,] -0.4352168 -0.4294538 -0.4009647 -0.4459887 -0.2952938 -0.4147294
#>              V58        V59         V6        V60         V7          V8
#>  [1,] -0.4333104 -0.2973847 -0.1354993 -0.3077434 -0.1988803 -0.21495163
#>  [2,] -0.3249798 -0.4072284 -0.1894053 -0.4151553 -0.1043180 -0.14177224
#>  [3,] -0.2436760 -0.3479907 -0.1137976 -0.1871920 -0.1713561 -0.08493956
#>  [4,] -0.2524635 -0.4112663 -0.1418151 -0.2432589 -0.1445189 -0.17377985
#>  [5,] -0.4030342 -0.3097421 -0.1325919 -0.4485266 -0.2957319 -0.26359251
#>  [6,] -0.4480525 -0.4395237 -0.3188737 -0.3754990 -0.2879881 -0.16494157
#>  [7,] -0.3005776 -0.2298694 -0.2201664 -0.1905378 -0.1923096 -0.22573596
#>  [8,] -0.2933784 -0.4079660 -0.1308902 -0.3183487 -0.1202917 -0.19929582
#>  [9,] -0.3320249 -0.3033636 -0.2908498 -0.3480528 -0.1838443 -0.08939803
#> [10,] -0.3866933 -0.3572935 -0.1943611 -0.4412482 -0.1862861 -0.19686882
#>                V9
#>  [1,] -0.25991932
#>  [2,] -0.03745566
#>  [3,] -0.18889460
#>  [4,] -0.20663522
#>  [5,] -0.23640782
#>  [6,] -0.18615034
#>  [7,] -0.06171990
#>  [8,] -0.24883280
#>  [9,] -0.02389384
#> [10,] -0.07195261
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.04661784  0.04843037 -0.05065013 -0.06805518  0.06155707  0.05851129
#> [2,] -0.02282771 -0.11603548  0.08353409  0.07196548 -0.06391015 -0.07670839
#>              [,7]        [,8]        [,9]      [,10]
#> [1,] -0.003161358 -0.03374694  0.13033213 0.10564685
#> [2,] -0.042725484  0.05075660 -0.02399239 0.02278993
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  4.720600e-05  0.0003392822  0.0003655662  0.0003849380  0.0003212535
#>  [2,] -1.668559e-04 -0.0011026440 -0.0011693200 -0.0011385906 -0.0009462854
#>  [3,]  2.836360e-04  0.0019598394  0.0021111594  0.0020559492  0.0016539666
#>  [4,]  2.164197e-04  0.0015418503  0.0017444963  0.0018324306  0.0013993902
#>  [5,] -1.727352e-04 -0.0012255324 -0.0013753828 -0.0013892969 -0.0010414631
#>  [6,] -5.990206e-05 -0.0003589208 -0.0003487615 -0.0003042665 -0.0002701114
#>  [7,] -3.658469e-05 -0.0002777636 -0.0003183733 -0.0003238181 -0.0002216710
#>  [8,]  1.053676e-04  0.0007773857  0.0008268723  0.0008422348  0.0006691829
#>  [9,] -1.038999e-04 -0.0006855280 -0.0007009094 -0.0006678833 -0.0005668928
#> [10,] -6.508098e-05 -0.0004275939 -0.0004447481 -0.0004236247 -0.0003412193
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0002611933  1.720143e-04  1.682621e-04  1.879621e-04  2.535317e-04
#>  [2,] -0.0007361546 -4.864007e-04 -4.324713e-04 -3.672855e-04 -4.953339e-04
#>  [3,]  0.0012715084  8.211451e-04  6.745879e-04  6.069931e-04  8.834562e-04
#>  [4,]  0.0009572533  5.372341e-04  3.724196e-04  3.889341e-04  6.949383e-04
#>  [5,] -0.0006924382 -3.539347e-04 -1.776604e-04 -1.312895e-04 -3.569238e-04
#>  [6,] -0.0002577632 -2.021046e-04 -2.112005e-04 -1.622296e-04 -1.718577e-04
#>  [7,] -0.0001256183 -5.104849e-05 -3.056461e-06  7.555843e-06 -3.865123e-05
#>  [8,]  0.0005110790  2.965167e-04  2.482428e-04  2.452895e-04  3.726774e-04
#>  [9,] -0.0004736330 -3.207874e-04 -3.126987e-04 -2.667557e-04 -3.020638e-04
#> [10,] -0.0002634950 -1.528745e-04 -1.133654e-04 -6.822392e-05 -1.126952e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0004324131  5.375026e-05  0.0005385997  0.0005570617  0.0004682300
#>  [2,] -0.0011311397 -1.888547e-04 -0.0016238548 -0.0016697573 -0.0012334432
#>  [3,]  0.0020051935  3.254163e-04  0.0028222637  0.0029280022  0.0022337881
#>  [4,]  0.0016381982  2.571527e-04  0.0021926516  0.0023216232  0.0018804655
#>  [5,] -0.0011558868 -1.989410e-04 -0.0016737341 -0.0017601773 -0.0013401377
#>  [6,] -0.0003551088 -6.202807e-05 -0.0005461516 -0.0005484785 -0.0003535067
#>  [7,] -0.0002052120 -4.455091e-05 -0.0003066322 -0.0003378818 -0.0002527928
#>  [8,]  0.0008028502  1.199154e-04  0.0010762533  0.0011212930  0.0008882059
#>  [9,] -0.0006604413 -1.174975e-04 -0.0009509867 -0.0009786088 -0.0007097756
#> [10,] -0.0003811582 -7.070346e-05 -0.0005901086 -0.0006006032 -0.0004168428
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0004234238  0.0004079578  0.0004409106  3.259342e-04  3.169407e-04
#>  [2,] -0.0010261898 -0.0009955786 -0.0010898609 -8.593651e-04 -8.997125e-04
#>  [3,]  0.0018416934  0.0016999533  0.0018098867  1.284066e-03  1.321220e-03
#>  [4,]  0.0015654877  0.0014276618  0.0014743579  9.429187e-04  1.016150e-03
#>  [5,] -0.0010469042 -0.0009027524 -0.0009203878 -5.188023e-04 -5.707968e-04
#>  [6,] -0.0002666348 -0.0002417533 -0.0002819649 -2.088537e-04 -1.745225e-04
#>  [7,] -0.0001743869 -0.0001336839 -0.0001324573 -3.601235e-05 -5.910245e-05
#>  [8,]  0.0007610625  0.0007198172  0.0007994114  5.578183e-04  5.466318e-04
#>  [9,] -0.0005846743 -0.0005621416 -0.0006445143 -5.032798e-04 -5.127841e-04
#> [10,] -0.0003240262 -0.0002895618 -0.0003229613 -2.154258e-04 -2.105869e-04
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0003861557  0.0003922909  3.180599e-05  3.515969e-04  1.395503e-04
#>  [2,] -0.0012569478 -0.0012242647 -1.247902e-04 -1.121839e-03 -4.024755e-04
#>  [3,]  0.0018727599  0.0017601976  2.137961e-04  1.531858e-03  2.834511e-04
#>  [4,]  0.0014641190  0.0013967484  1.578471e-04  1.096144e-03  6.845430e-05
#>  [5,] -0.0009345244 -0.0008332921 -1.235146e-04 -6.041147e-04  2.105768e-04
#>  [6,] -0.0002975362 -0.0003417631 -4.020120e-05 -3.657919e-04 -9.764365e-05
#>  [7,] -0.0001522630 -0.0001266036 -2.661330e-05 -6.844038e-05  1.178968e-04
#>  [8,]  0.0007331219  0.0007310683  6.883139e-05  6.416025e-04  1.496394e-04
#>  [9,] -0.0007171916 -0.0007068531 -7.729264e-05 -6.879035e-04 -2.533319e-04
#> [10,] -0.0003556664 -0.0003570400 -4.371781e-05 -3.317591e-04 -3.957890e-05
#>                 V32           V33           V34           V35           V36
#>  [1,]  1.245961e-04  5.307300e-05 -6.264533e-05 -2.204396e-04 -0.0002931838
#>  [2,] -4.462324e-04 -2.604573e-04 -4.648220e-05  4.015220e-04  0.0006910510
#>  [3,]  4.034252e-04  4.951234e-05 -4.115343e-04 -1.252973e-03 -0.0017300405
#>  [4,]  1.440409e-04 -2.358665e-04 -7.173075e-04 -1.475525e-03 -0.0018662245
#>  [5,]  6.471917e-05  3.303052e-04  6.741123e-04  1.283919e-03  0.0016051055
#>  [6,] -1.271489e-04 -1.256188e-04 -9.325411e-05  2.143675e-05  0.0001203989
#>  [7,]  8.144703e-05  1.525874e-04  2.442547e-04  3.787151e-04  0.0004410287
#>  [8,]  1.474208e-04 -1.806521e-05 -2.857169e-04 -6.392508e-04 -0.0007920118
#>  [9,] -2.626038e-04 -1.384911e-04  1.066265e-05  2.757238e-04  0.0004439143
#> [10,] -8.753321e-05 -3.216071e-05  4.875823e-05  2.372063e-04  0.0003660068
#>                 V37           V38           V39            V4           V40
#>  [1,] -1.781594e-04  9.885816e-05  0.0002205101  6.692992e-05  1.555188e-04
#>  [2,]  3.608301e-04 -4.857780e-04 -0.0008023463 -2.480039e-04 -4.666146e-04
#>  [3,] -1.061969e-03  5.658361e-04  0.0012466961  4.375939e-04  6.912928e-04
#>  [4,] -1.219673e-03  2.256810e-04  0.0008875093  3.380592e-04  4.606634e-04
#>  [5,]  1.077646e-03 -1.104712e-04 -0.0006600557 -2.718812e-04 -2.821977e-04
#>  [6,]  4.168779e-05 -2.104253e-04 -0.0002639786 -8.024028e-05 -1.635535e-04
#>  [7,]  2.964344e-04  1.157406e-05 -0.0001220851 -5.934242e-05 -3.118159e-05
#>  [8,] -4.987615e-04  1.524683e-04  0.0004441364  1.491591e-04  2.808151e-04
#>  [9,]  2.303156e-04 -2.846582e-04 -0.0004683013 -1.492006e-04 -2.880405e-04
#> [10,]  2.308507e-04 -1.328833e-04 -0.0002730767 -9.427997e-05 -1.457087e-04
#>                 V41           V42           V43           V44           V45
#>  [1,]  2.522685e-04  0.0003191221  0.0002753747  0.0002794447  0.0003165179
#>  [2,] -7.284568e-04 -0.0009376205 -0.0008879482 -0.0009288825 -0.0011528155
#>  [3,]  1.177339e-03  0.0015739162  0.0014909718  0.0016001184  0.0019401026
#>  [4,]  8.914381e-04  0.0012773380  0.0011901784  0.0012693202  0.0014836043
#>  [5,] -6.070099e-04 -0.0009130395 -0.0008885500 -0.0009839591 -0.0011910943
#>  [6,] -2.500476e-04 -0.0002979242 -0.0002771801 -0.0002907663 -0.0003765151
#>  [7,] -9.856294e-05 -0.0001768541 -0.0001758493 -0.0002034795 -0.0002528418
#>  [8,]  4.966039e-04  0.0006632499  0.0005852716  0.0006161591  0.0007347932
#>  [9,] -4.799532e-04 -0.0005963043 -0.0005395523 -0.0005546968 -0.0007163998
#> [10,] -2.648247e-04 -0.0003503744 -0.0003339895 -0.0003571451 -0.0004496473
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0002809711  0.0002248192  0.0001561353  9.213407e-05  8.259315e-05
#>  [2,] -0.0010173174 -0.0007340587 -0.0005212095 -3.102884e-04 -2.623989e-04
#>  [3,]  0.0016832280  0.0012509900  0.0008855101  5.339384e-04  4.597011e-04
#>  [4,]  0.0012963753  0.0010230288  0.0006969010  4.210496e-04  3.670575e-04
#>  [5,] -0.0010232014 -0.0007838441 -0.0005423845 -3.333679e-04 -2.805466e-04
#>  [6,] -0.0003264609 -0.0002154945 -0.0001582799 -9.565570e-05 -8.354955e-05
#>  [7,] -0.0002201201 -0.0001723647 -0.0001163115 -7.170926e-05 -5.696437e-05
#>  [8,]  0.0006549809  0.0005097298  0.0003544138  2.114978e-04  1.751738e-04
#>  [9,] -0.0006503247 -0.0004686864 -0.0003335433 -1.985006e-04 -1.626849e-04
#> [10,] -0.0003910702 -0.0002783555 -0.0001979445 -1.197436e-04 -1.001843e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.395976e-05  2.279087e-05  2.234326e-05  1.427422e-05  1.108380e-05
#>  [2,] -8.959872e-05 -7.800943e-05 -7.476313e-05 -4.679148e-05 -4.193369e-05
#>  [3,]  1.402288e-04  1.290442e-04  1.285261e-04  7.906723e-05  6.740013e-05
#>  [4,]  9.929987e-05  1.068868e-04  1.043118e-04  6.248111e-05  4.825390e-05
#>  [5,] -7.740663e-05 -7.866492e-05 -7.929952e-05 -4.725515e-05 -3.781421e-05
#>  [6,] -3.133919e-05 -2.279142e-05 -2.334784e-05 -1.579821e-05 -1.539829e-05
#>  [7,] -1.399484e-05 -1.753678e-05 -1.750898e-05 -9.071949e-06 -6.830146e-06
#>  [8,]  5.176688e-05  4.989453e-05  4.906818e-05  2.984239e-05  2.292133e-05
#>  [9,] -5.766949e-05 -4.891585e-05 -4.706397e-05 -2.821305e-05 -2.485499e-05
#> [10,] -3.370003e-05 -2.838226e-05 -2.819276e-05 -1.820738e-05 -1.574538e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  9.067731e-06  1.194897e-05  1.803161e-06  9.398962e-06  1.081382e-05
#>  [2,] -3.492062e-05 -4.049122e-05 -1.540914e-05 -3.666807e-05 -3.381275e-05
#>  [3,]  5.606277e-05  6.704033e-05  1.873879e-05  5.958672e-05  5.680393e-05
#>  [4,]  3.798929e-05  5.063250e-05  7.676287e-06  4.549038e-05  4.545115e-05
#>  [5,] -2.969804e-05 -3.955368e-05 -6.475947e-06 -3.574992e-05 -3.368242e-05
#>  [6,] -1.209616e-05 -1.501081e-05 -5.898013e-06 -1.220585e-05 -1.174071e-05
#>  [7,] -4.876297e-06 -7.568041e-06 -1.698966e-07 -7.385499e-06 -6.653540e-06
#>  [8,]  1.754987e-05  2.515825e-05  3.003158e-06  1.946267e-05  2.178966e-05
#>  [9,] -2.129818e-05 -2.529427e-05 -9.089641e-06 -2.239471e-05 -2.053678e-05
#> [10,] -1.273173e-05 -1.615622e-05 -5.039877e-06 -1.382964e-05 -1.286787e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  9.142446e-05  4.044513e-06  1.199355e-04  0.0001989560  0.0003226322
#>  [2,] -2.545464e-04 -1.550854e-05 -3.539367e-04 -0.0006066508 -0.0010389984
#>  [3,]  4.471709e-04  2.367635e-05  6.032939e-04  0.0010743453  0.0018544869
#>  [4,]  3.799620e-04  1.785393e-05  4.901519e-04  0.0008506920  0.0014412153
#>  [5,] -2.654731e-04 -1.207551e-05 -3.498696e-04 -0.0006548719 -0.0011535260
#>  [6,] -7.135668e-05 -5.238884e-06 -1.095433e-04 -0.0002111438 -0.0003490101
#>  [7,] -5.559467e-05 -2.447360e-06 -7.218525e-05 -0.0001457647 -0.0002571949
#>  [8,]  1.799302e-04  7.631533e-06  2.425766e-04  0.0004436646  0.0007437067
#>  [9,] -1.523480e-04 -9.270464e-06 -2.127730e-04 -0.0003784108 -0.0006480527
#> [10,] -8.650579e-05 -5.149305e-06 -1.256119e-04 -0.0002348175 -0.0004065380
#> 
#> $vW[[2]]
#>              [,1]        [,2]        [,3]      [,4]         [,5]        [,6]
#> [1,] -0.007977863 -0.02499305 -0.01782513 -0.010822 -0.007964491 -0.03436961
#> [2,]  0.007977863  0.02499305  0.01782513  0.010822  0.007964491  0.03436961
#>              [,7]          [,8]        [,9]       [,10]
#> [1,] -0.003045764 -0.0007432537 -0.03104713 -0.02088148
#> [2,]  0.003045764  0.0007432537  0.03104713  0.02088148
#> 
#> 
#> $B
#> $B[[1]]
#>  [1]  0.31733423  0.55235409  0.23477237  0.26245504  0.32396991  0.80738654
#>  [7] -0.07898548  0.11802957  0.60342334  0.51460633
#> 
#> $B[[2]]
#> [1]  0.006764653 -0.119820068
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  4.756791e-04 -1.393758e-03  2.028715e-03  1.410603e-03 -7.920098e-04
#>  [6] -4.417114e-04 -5.557011e-05  7.945714e-04 -8.254248e-04 -3.873887e-04
#> 
#> $vB[[2]]
#> [1] -0.04674458  0.04674458
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#>  [2,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#>  [3,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#>  [4,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#>  [5,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#>  [6,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#>  [7,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#>  [8,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#>  [9,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [10,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [11,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [12,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [13,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [14,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [15,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [16,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [17,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#> [18,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [19,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [20,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [21,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [22,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [23,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [24,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [25,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [26,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [27,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [28,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [29,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [30,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [31,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [32,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [33,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [34,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [35,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [36,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [37,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [38,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [39,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [40,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#>  [2,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#>  [3,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#>  [4,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#>  [5,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#>  [6,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#>  [7,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#>  [8,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#>  [9,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [10,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [11,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [12,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [13,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [14,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [15,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [16,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [17,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#> [18,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [19,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [20,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [21,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [22,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [23,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [24,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [25,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [26,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [27,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [28,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [29,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [30,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [31,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [32,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [33,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [34,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [35,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [36,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [37,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [38,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [39,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [40,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#>  [2,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#>  [3,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#>  [4,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#>  [5,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#>  [6,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#>  [7,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#>  [8,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#>  [9,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [10,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [11,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [12,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [13,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [14,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [15,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [16,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [17,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#> [18,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [19,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [20,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [21,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [22,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [23,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [24,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [25,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [26,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [27,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [28,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [29,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [30,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [31,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [32,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [33,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [34,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [35,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [36,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [37,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [38,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [39,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [40,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#>  [2,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#>  [3,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#>  [4,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#>  [5,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#>  [6,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#>  [7,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#>  [8,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#>  [9,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [10,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [11,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [12,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [13,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [14,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [15,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [16,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [17,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#> [18,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [19,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [20,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [21,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [22,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [23,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [24,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [25,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [26,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [27,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [28,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [29,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [30,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [31,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [32,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [33,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [34,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [35,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [36,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [37,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [38,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [39,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [40,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#>  [2,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#>  [3,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#>  [4,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#>  [5,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#>  [6,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#>  [7,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#>  [8,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#>  [9,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [10,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [11,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [12,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [13,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [14,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [15,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [16,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [17,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#> [18,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [19,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [20,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [21,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [22,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [23,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [24,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [25,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [26,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [27,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [28,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [29,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [30,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [31,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [32,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [33,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [34,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [35,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [36,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [37,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [38,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [39,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [40,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#>  [2,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#>  [3,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#>  [4,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#>  [5,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#>  [6,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#>  [7,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#>  [8,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#>  [9,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [10,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [11,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [12,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [13,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [14,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [15,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [16,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [17,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#> [18,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [19,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [20,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [21,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [22,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [23,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [24,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [25,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [26,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [27,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [28,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [29,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [30,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [31,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [32,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [33,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [34,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [35,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [36,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [37,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [38,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [39,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [40,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4781761 0.7765019 0.5656776 0.3514894 0.4419818 0.9021729 0.2518290
#>  [2,] 0.5948503 0.8040871 0.6931011 0.5229757 0.6395287 0.9305054 0.3398126
#>  [3,] 0.4479512 0.7931507 0.5679856 0.3954232 0.4753173 0.8754589 0.2486831
#>  [4,] 0.5798858 0.7988600 0.6734154 0.4949738 0.6002772 0.9344433 0.3239517
#>  [5,] 0.5096378 0.7714367 0.6420671 0.4568548 0.5374982 0.8981752 0.2546216
#>  [6,] 0.5683354 0.7884689 0.6673665 0.4892766 0.5925936 0.9313747 0.3112537
#>  [7,] 0.4784874 0.7879769 0.6071363 0.4125745 0.4430406 0.8737462 0.2640019
#>  [8,] 0.4638851 0.7866162 0.5868567 0.3769592 0.5091991 0.8630085 0.2643828
#>  [9,] 0.3242366 0.7011933 0.4905734 0.2664061 0.3247792 0.8679888 0.1480793
#> [10,] 0.4028999 0.7038363 0.5406236 0.3541195 0.4329342 0.8462612 0.2004047
#> [11,] 0.4258758 0.7796545 0.5728311 0.3609730 0.4618542 0.8774211 0.2160185
#> [12,] 0.4744645 0.8070828 0.5990271 0.3462298 0.5364742 0.9236590 0.2628073
#> [13,] 0.3676047 0.7386111 0.5354417 0.3483908 0.3941128 0.7917627 0.2071169
#> [14,] 0.5318146 0.7546519 0.6292658 0.4626009 0.5282213 0.9117787 0.2819953
#> [15,] 0.5820227 0.7796348 0.6673384 0.4843005 0.5964897 0.9314520 0.3108225
#> [16,] 0.4034781 0.8396271 0.5887492 0.2716760 0.4086729 0.9240389 0.1514100
#> [17,] 0.2579368 0.6875694 0.3694206 0.2165167 0.2744928 0.7019781 0.1585500
#> [18,] 0.2264299 0.7120983 0.3600196 0.1689366 0.2241147 0.7319947 0.1204711
#> [19,] 0.3261274 0.7404845 0.4486936 0.2548534 0.3470368 0.7867353 0.1901794
#> [20,] 0.3760057 0.6913898 0.4882063 0.2945150 0.3519012 0.8310535 0.1927819
#> [21,] 0.5412534 0.7977593 0.6444129 0.4230578 0.5774182 0.9386027 0.2859907
#> [22,] 0.3118906 0.7051890 0.4497404 0.2216620 0.2904957 0.8644084 0.1339241
#> [23,] 0.4196979 0.7583132 0.5786108 0.3768532 0.4625873 0.8678173 0.2064379
#> [24,] 0.6155601 0.8294992 0.7143468 0.5434868 0.6302260 0.9381323 0.3215712
#> [25,] 0.4852495 0.7726299 0.6190358 0.4036409 0.5061247 0.9232569 0.2471574
#> [26,] 0.4819241 0.7983929 0.5788462 0.3636445 0.4516991 0.9012769 0.2448612
#> [27,] 0.4530886 0.7256057 0.5732895 0.3919362 0.4376995 0.8854380 0.2409265
#> [28,] 0.5508738 0.7986248 0.6786732 0.4805972 0.5568520 0.9373099 0.2571615
#> [29,] 0.4946662 0.7820588 0.5994736 0.4451841 0.5328387 0.8893940 0.3155499
#> [30,] 0.4369225 0.7834275 0.5832477 0.3637783 0.4571654 0.8680263 0.2259432
#> [31,] 0.5123395 0.7539735 0.6194164 0.4462046 0.5004669 0.9055812 0.2681169
#> [32,] 0.4873275 0.7518971 0.5731930 0.4565852 0.5215837 0.8737844 0.2700054
#> [33,] 0.4277409 0.8147551 0.6122996 0.3381228 0.5075972 0.9285365 0.2375360
#> [34,] 0.4927820 0.7878034 0.6114002 0.4168269 0.5120907 0.8719145 0.2801577
#> [35,] 0.4369903 0.7336730 0.5922914 0.3677378 0.4514749 0.8883215 0.1913383
#> [36,] 0.4898166 0.8087773 0.6236520 0.3787774 0.4914346 0.9277265 0.2166723
#> [37,] 0.4599336 0.7738957 0.5702252 0.3636974 0.5028707 0.9025663 0.2483381
#> [38,] 0.3230733 0.6821140 0.4715757 0.2874924 0.3030393 0.8343037 0.1456228
#> [39,] 0.4001242 0.7810909 0.5589957 0.3153786 0.4541134 0.9010216 0.2144787
#> [40,] 0.5014706 0.7920084 0.6386608 0.4201693 0.5463645 0.9120663 0.2536649
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5076496 0.7943770 0.7354462
#>  [2,] 0.4674434 0.8691898 0.7916280
#>  [3,] 0.3954877 0.8428124 0.7189667
#>  [4,] 0.5052834 0.8555585 0.7911385
#>  [5,] 0.3593392 0.8573059 0.7281946
#>  [6,] 0.4862947 0.8521201 0.7839883
#>  [7,] 0.3924453 0.8238752 0.6849681
#>  [8,] 0.3671517 0.8506291 0.7117971
#>  [9,] 0.2684900 0.7761911 0.6218037
#> [10,] 0.3514486 0.8163636 0.6514574
#> [11,] 0.3318487 0.8457970 0.7064788
#> [12,] 0.5281801 0.8478280 0.7895737
#> [13,] 0.2676151 0.7995204 0.5925809
#> [14,] 0.5009875 0.8095903 0.7335371
#> [15,] 0.5199378 0.8470987 0.7875427
#> [16,] 0.4146239 0.8828531 0.7487572
#> [17,] 0.2824631 0.7542937 0.5422750
#> [18,] 0.2678480 0.7679134 0.5411988
#> [19,] 0.4339407 0.7594118 0.6287979
#> [20,] 0.4075534 0.7256093 0.6372942
#> [21,] 0.5751600 0.8440991 0.7989979
#> [22,] 0.3651249 0.7697844 0.6362061
#> [23,] 0.3118119 0.8464214 0.6851110
#> [24,] 0.4805179 0.8899410 0.8011737
#> [25,] 0.4763663 0.8193057 0.7395646
#> [26,] 0.5033717 0.8214717 0.7447506
#> [27,] 0.4516386 0.7727666 0.6758935
#> [28,] 0.4400628 0.8742478 0.7737960
#> [29,] 0.4556950 0.8282415 0.7314114
#> [30,] 0.3577979 0.8437806 0.7119446
#> [31,] 0.4796646 0.8032830 0.7170107
#> [32,] 0.3852036 0.8176212 0.7116430
#> [33,] 0.4130793 0.8637162 0.7588961
#> [34,] 0.3973918 0.8324756 0.7287897
#> [35,] 0.3392013 0.8374779 0.6782820
#> [36,] 0.4778568 0.8672506 0.7624461
#> [37,] 0.5140177 0.8222107 0.7530882
#> [38,] 0.3137172 0.7287718 0.6065890
#> [39,] 0.3778607 0.8514876 0.7234985
#> [40,] 0.4071242 0.8660012 0.7645238
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5316437 0.4683563
#>  [2,] 0.5287269 0.4712731
#>  [3,] 0.5343364 0.4656636
#>  [4,] 0.5284196 0.4715804
#>  [5,] 0.5308248 0.4691752
#>  [6,] 0.5286112 0.4713888
#>  [7,] 0.5300426 0.4699574
#>  [8,] 0.5354983 0.4645017
#>  [9,] 0.5365313 0.4634687
#> [10,] 0.5326599 0.4673401
#> [11,] 0.5360467 0.4639533
#> [12,] 0.5357176 0.4642824
#> [13,] 0.5333507 0.4666493
#> [14,] 0.5265602 0.4734398
#> [15,] 0.5277294 0.4722706
#> [16,] 0.5391881 0.4608119
#> [17,] 0.5382798 0.4617202
#> [18,] 0.5410175 0.4589825
#> [19,] 0.5356264 0.4643736
#> [20,] 0.5312315 0.4687685
#> [21,] 0.5303431 0.4696569
#> [22,] 0.5371133 0.4628867
#> [23,] 0.5346232 0.4653768
#> [24,] 0.5275360 0.4724640
#> [25,] 0.5304514 0.4695486
#> [26,] 0.5321856 0.4678144
#> [27,] 0.5280985 0.4719015
#> [28,] 0.5290114 0.4709886
#> [29,] 0.5310846 0.4689154
#> [30,] 0.5347462 0.4652538
#> [31,] 0.5269780 0.4730220
#> [32,] 0.5310680 0.4689320
#> [33,] 0.5377588 0.4622412
#> [34,] 0.5320727 0.4679273
#> [35,] 0.5325879 0.4674121
#> [36,] 0.5329647 0.4670353
#> [37,] 0.5333873 0.4666127
#> [38,] 0.5325725 0.4674275
#> [39,] 0.5382460 0.4617540
#> [40,] 0.5331247 0.4668753
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]      [,2]        [,3]        [,4]         [,5]      [,6]
#>  [1,] -0.087351150 1.2453960  0.26423706 -0.61249863 -0.233122883 2.2216043
#>  [2,]  0.384053291 1.4120375  0.81465741  0.09196758  0.573319249 2.5944789
#>  [3,] -0.208952156 1.3440224  0.27363710 -0.42457232 -0.098811251 1.9501122
#>  [4,]  0.322304570 1.3791847  0.72367329 -0.02010556  0.406620060 2.6570352
#>  [5,]  0.038556116 1.2164416  0.58434741 -0.17301099  0.150274861 2.1771120
#>  [6,]  0.275062756 1.3157210  0.69629829 -0.04290031  0.374697677 2.6080003
#>  [7,] -0.086103608 1.3127734  0.43529055 -0.35333269 -0.228830788 1.9344961
#>  [8,] -0.144711519 1.3046481  0.35098630 -0.50247501  0.036800669 1.8405055
#>  [9,] -0.734369349 0.8529869 -0.03771080 -1.01293379 -0.731894074 1.8832919
#> [10,] -0.393396736 0.8656333  0.16285318 -0.60098021 -0.269889453 1.7055732
#> [11,] -0.298697871 1.2636541  0.29341124 -0.57114372 -0.152880336 1.9682323
#> [12,] -0.102231030 1.4311652  0.40141291 -0.63565327  0.146156559 2.4931326
#> [13,] -0.542506304 1.0387620  0.14200514 -0.62612037 -0.430056780 1.3355832
#> [14,]  0.127430454 1.1235785  0.52906836 -0.14987624  0.113005297 2.3355494
#> [15,]  0.331082509 1.2635395  0.69617134 -0.06281855  0.390860041 2.6092105
#> [16,] -0.390993619 1.6554560  0.35879686 -0.98613565 -0.369454304 2.4985329
#> [17,] -1.056719837 0.7887806 -0.53470313 -1.28608225 -0.971945724 0.8567354
#> [18,] -1.228580931 0.9055971 -0.57527897 -1.59318255 -1.241846899 1.0047666
#> [19,] -0.725753175 1.0484883 -0.20595057 -1.07289272 -0.632089917 1.3053578
#> [20,] -0.506536799 0.8066247 -0.04718372 -0.87355565 -0.610692591 1.5931124
#> [21,]  0.165389401 1.3723484  0.59456942 -0.31023342  0.312183898 2.7270259
#> [22,] -0.791295170 0.8721312 -0.20171967 -1.25600743 -0.892977586 1.8523977
#> [23,] -0.324013848 1.1434541  0.31707315 -0.50292655 -0.149931153 1.8817964
#> [24,]  0.470745104 1.5820825  0.91659011  0.17438773  0.533186372 2.7188932
#> [25,] -0.059019092 1.2232212  0.48545765 -0.39031753  0.024500071 2.4874439
#> [26,] -0.072334948 1.3762803  0.31803890 -0.55958059 -0.193807966 2.2114936
#> [27,] -0.188199318 0.9724406  0.29528521 -0.43918074 -0.250503783 2.0449659
#> [28,]  0.204201982 1.3777214  0.74768094 -0.07765000  0.228395655 2.7048110
#> [29,] -0.021335973 1.2777044  0.40327216 -0.22014855  0.131544174 2.0845661
#> [30,] -0.253661368 1.2857534  0.33611973 -0.55900234 -0.171759470 1.8836194
#> [31,]  0.049368079 1.1199182  0.48707186 -0.21601767  0.001867687 2.2608362
#> [32,] -0.050700804 1.1087558  0.29489075 -0.17409778  0.086388594 1.9348421
#> [33,] -0.291074130 1.4812086  0.45698888 -0.67167089  0.030390985 2.5644232
#> [34,] -0.028873837 1.3117352  0.45320162 -0.33581309  0.048372237 1.9179937
#> [35,] -0.253385768 1.0133387  0.37344590 -0.54193394 -0.194713120 2.0737097
#> [36,] -0.040739178 1.4420850  0.50507824 -0.49474087 -0.034265131 2.5522800
#> [37,] -0.160609777 1.2304408  0.28277020 -0.55935202  0.011482932 2.2260703
#> [38,] -0.739683768 0.7635038 -0.11381992 -0.90759420 -0.832866440 1.6164407
#> [39,] -0.404947768 1.2720351  0.23708703 -0.77509225 -0.184064271 2.2086274
#> [40,]  0.005882264 1.3370740  0.56955651 -0.32207825  0.185992489 2.3391295
#>             [,7]         [,8]      [,9]     [,10]
#>  [1,] -1.0888813  0.030600867 1.3515135 1.0224327
#>  [2,] -0.6641295 -0.130410882 1.8938142 1.3347669
#>  [3,] -1.1056482 -0.424302407 1.6793044 0.9393421
#>  [4,] -0.7356703  0.021134502 1.7788800 1.3318016
#>  [5,] -1.0741136 -0.578233146 1.7930919 0.9854822
#>  [6,] -0.7942646 -0.054834834 1.7513268 1.2890617
#>  [7,] -1.0252711 -0.437044936 1.5428264 0.7766984
#>  [8,] -1.0233120 -0.544455459 1.7395434 0.9041281
#>  [9,] -1.7497452 -1.002297334 1.2436064 0.4972110
#> [10,] -1.3837670 -0.612677953 1.4919023 0.6254516
#> [11,] -1.2890214 -0.699835605 1.7020093 0.8783435
#> [12,] -1.0314284  0.112840024 1.7176667 1.3223577
#> [13,] -1.3423925 -1.006756570 1.3832998 0.3746448
#> [14,] -0.9345855  0.003949899 1.4473500 1.0126432
#> [15,] -0.7962769  0.079793427 1.7120248 1.3101767
#> [16,] -1.7235851 -0.344882779 2.0197303 1.0919948
#> [17,] -1.6690564 -0.932276668 1.1216446 0.1695049
#> [18,] -1.9879770 -1.005568461 1.1965663 0.1651695
#> [19,] -1.4488450 -0.265790858 1.1494573 0.5270633
#> [20,] -1.4320345 -0.374088731 0.9724586 0.5636393
#> [21,] -0.9149365  0.302935624 1.6890495 1.3800428
#> [22,] -1.8666997 -0.553188626 1.2070939 0.5589352
#> [23,] -1.3465321 -0.791662131 1.7068051 0.7773608
#> [24,] -0.7465606 -0.077968010 2.0901386 1.3936459
#> [25,] -1.1138307 -0.094605215 1.5116506 1.0437070
#> [26,] -1.1262100  0.013487029 1.5263503 1.0708081
#> [27,] -1.1476069 -0.194052069 1.2239996 0.7349636
#> [28,] -1.0607743 -0.240907219 1.9390501 1.2298711
#> [29,] -0.7742988 -0.177686085 1.5732156 1.0017950
#> [30,] -1.2313614 -0.584934772 1.6866310 0.9048474
#> [31,] -1.0041980 -0.081386455 1.4069411 0.9296815
#> [32,] -0.9945950 -0.467518980 1.5003135 0.9033770
#> [33,] -1.1662363 -0.351249927 1.8465048 1.1466367
#> [34,] -0.9436796 -0.416344691 1.6032753 0.9884905
#> [35,] -1.4413376 -0.666855529 1.6395808 0.7458876
#> [36,] -1.2851652 -0.088630719 1.8768647 1.1661375
#> [37,] -1.1074955  0.056085328 1.5313977 1.1151511
#> [38,] -1.7693529 -0.782797866 0.9884001 0.4329966
#> [39,] -1.2981369 -0.498638203 1.7463163 0.9618823
#> [40,] -1.0791605 -0.375866707 1.8660556 1.1776430
#> 
#> $pre[[3]]
#>             [,1]        [,2]
#>  [1,] 0.05325835 -0.07348577
#>  [2,] 0.04958496 -0.06544924
#>  [3,] 0.06011694 -0.07744500
#>  [4,] 0.04922608 -0.06457487
#>  [5,] 0.05363384 -0.06982209
#>  [6,] 0.04977713 -0.06479288
#>  [7,] 0.04777146 -0.07254386
#>  [8,] 0.06182149 -0.08041115
#>  [9,] 0.06209933 -0.08428672
#> [10,] 0.05610867 -0.07471723
#> [11,] 0.06435270 -0.08008483
#> [12,] 0.06701056 -0.07610350
#> [13,] 0.05244793 -0.08115307
#> [14,] 0.04294429 -0.06339669
#> [15,] 0.04807297 -0.06295839
#> [16,] 0.07601887 -0.08105564
#> [17,] 0.06146577 -0.09195348
#> [18,] 0.06818372 -0.09625574
#> [19,] 0.05888639 -0.08386109
#> [20,] 0.04876723 -0.07632177
#> [21,] 0.05519731 -0.06632434
#> [22,] 0.06460822 -0.08411847
#> [23,] 0.06135018 -0.07736451
#> [24,] 0.04819449 -0.06206091
#> [25,] 0.05289285 -0.06906365
#> [26,] 0.05587503 -0.07304545
#> [27,] 0.04389600 -0.06861653
#> [28,] 0.05236995 -0.06380599
#> [29,] 0.05176112 -0.07273768
#> [30,] 0.06173189 -0.07747719
#> [31,] 0.04311683 -0.06489999
#> [32,] 0.05135898 -0.07307339
#> [33,] 0.07114582 -0.08017755
#> [34,] 0.05431349 -0.07415350
#> [35,] 0.05719249 -0.07334405
#> [36,] 0.06129397 -0.07075639
#> [37,] 0.06025252 -0.07349584
#> [38,] 0.05183266 -0.07864220
#> [39,] 0.07092120 -0.08236221
#> [40,] 0.06079989 -0.07189346
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5316437  0.5316437
#>  [2,] -0.5287269  0.5287269
#>  [3,] -0.5343364  0.5343364
#>  [4,]  0.4715804 -0.4715804
#>  [5,]  0.4691752 -0.4691752
#>  [6,]  0.4713888 -0.4713888
#>  [7,] -0.5300426  0.5300426
#>  [8,] -0.5354983  0.5354983
#>  [9,]  0.4634687 -0.4634687
#> [10,] -0.5326599  0.5326599
#> [11,]  0.4639533 -0.4639533
#> [12,] -0.5357176  0.5357176
#> [13,]  0.4666493 -0.4666493
#> [14,]  0.4734398 -0.4734398
#> [15,]  0.4722706 -0.4722706
#> [16,] -0.5391881  0.5391881
#> [17,]  0.4617202 -0.4617202
#> [18,]  0.4589825 -0.4589825
#> [19,] -0.5356264  0.5356264
#> [20,]  0.4687685 -0.4687685
#> [21,] -0.5303431  0.5303431
#> [22,]  0.4628867 -0.4628867
#> [23,]  0.4653768 -0.4653768
#> [24,] -0.5275360  0.5275360
#> [25,]  0.4695486 -0.4695486
#> [26,] -0.5321856  0.5321856
#> [27,]  0.4719015 -0.4719015
#> [28,]  0.4709886 -0.4709886
#> [29,] -0.5310846  0.5310846
#> [30,]  0.4652538 -0.4652538
#> [31,]  0.4730220 -0.4730220
#> [32,] -0.5310680  0.5310680
#> [33,]  0.4622412 -0.4622412
#> [34,]  0.4679273 -0.4679273
#> [35,] -0.5325879  0.5325879
#> [36,] -0.5329647  0.5329647
#> [37,] -0.5333873  0.5333873
#> [38,]  0.4674275 -0.4674275
#> [39,]  0.4617540 -0.4617540
#> [40,]  0.4668753 -0.4668753
#> 
#> $L
#> [1] 0.6945098 0.6904904 0.7111652 0.6908908 0.6925493 0.6853602
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536