Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11          V12          V13         V14
#>  [1,] -0.4046306 -0.012644208 -0.04342730 -0.008317890 -0.059509125 -0.06146471
#>  [2,] -0.1892494 -0.133816723 -0.10323746 -0.112800490 -0.055600367 -0.07406103
#>  [3,] -0.2797758 -0.005216725 -0.13630664 -0.177932992  0.015514060  0.06707367
#>  [4,] -0.3562573 -0.075857128 -0.10131695 -0.060488325 -0.062690637 -0.00282991
#>  [5,] -0.2071473 -0.160450014 -0.17243603 -0.046271450 -0.048911839 -0.12729515
#>  [6,] -0.4242624 -0.119148062 -0.13875229 -0.168641672 -0.255919416 -0.13722190
#>  [7,] -0.1736276 -0.119069124 -0.05046514 -0.045173390 -0.076457613 -0.10479149
#>  [8,] -0.3627421 -0.178785113 -0.16089850 -0.009481939 -0.053171304  0.07080360
#>  [9,] -0.3518230 -0.198154947 -0.13442915 -0.083655345  0.008207085 -0.08792816
#> [10,] -0.2643360 -0.128852609 -0.05680582 -0.127257031 -0.005369085 -0.09629258
#>               V15         V16         V17         V18          V19         V2
#>  [1,]  0.01841787  0.05803949  0.13553911  0.08628939  0.206696460 -0.3256848
#>  [2,] -0.03597240  0.03156847 -0.06633255  0.01896737  0.022813079 -0.2308155
#>  [3,] -0.07087158  0.03496687  0.17367104  0.16998529  0.156592528 -0.2923271
#>  [4,]  0.03718258  0.01565143  0.02514996  0.07084650  0.178325988 -0.3514926
#>  [5,] -0.05842908 -0.07245736 -0.01759722 -0.05877583 -0.037490436 -0.3200764
#>  [6,] -0.12874152 -0.28011925 -0.26881591 -0.20955234 -0.129929291 -0.3071581
#>  [7,]  0.02379517 -0.05030945 -0.06541835  0.12374031  0.078730703 -0.2789699
#>  [8,] -0.11933380  0.02556892  0.12476092 -0.05388072  0.147993388 -0.3167310
#>  [9,] -0.09004312 -0.10065455 -0.04101512  0.02713157 -0.004129952 -0.2831726
#> [10,] -0.01891361 -0.01360351  0.08277764  0.10766865  0.048299748 -0.1968773
#>               V20          V21         V22          V23        V24          V25
#>  [1,]  0.06397519  0.105459808  0.19898579  0.307098825 0.36825148  0.193843960
#>  [2,]  0.18591126  0.133342061  0.26849346  0.258622505 0.14980640  0.237310329
#>  [3,]  0.07640383  0.131436951  0.08686450  0.126611649 0.11610835  0.163128857
#>  [4,]  0.22910163  0.104952564  0.13301863  0.172866559 0.12344984  0.151731958
#>  [5,] -0.06500527 -0.032599682 -0.03870018  0.007476073 0.03464911  0.002345310
#>  [6,] -0.04938643 -0.023178792 -0.17851515 -0.146385769 0.05320150  0.058071758
#>  [7,]  0.14610563  0.222040716  0.13357891  0.221818304 0.23726932  0.181252356
#>  [8,]  0.11056292  0.117540096  0.08446805  0.099359552 0.22663104  0.224401197
#>  [9,] -0.01367527 -0.009313915  0.18083127  0.042476493 0.03310890 -0.003695995
#> [10,]  0.01922384  0.135328871  0.17944442  0.118533565 0.11954772  0.211382181
#>              V26         V27         V28           V29         V3          V30
#>  [1,] 0.14421933  0.28286061  0.23764896  0.0902488925 -0.3047689  0.171883093
#>  [2,] 0.22301414  0.19067539  0.09050881  0.1993600677 -0.2675006  0.070412607
#>  [3,] 0.03791734  0.08500135 -0.02157061 -0.0004922895 -0.1529816 -0.029566221
#>  [4,] 0.12986060  0.17045073  0.08028724  0.0247002587 -0.2439561  0.032697230
#>  [5,] 0.18396642  0.17504348  0.06804247  0.0647489566 -0.3652766  0.057738261
#>  [6,] 0.11087091 -0.01243035  0.05670169  0.0554464214 -0.3363773  0.085471526
#>  [7,] 0.29772806  0.12115167  0.15952270  0.0388202404 -0.1952719  0.166194475
#>  [8,] 0.10529310  0.10097330  0.14461488  0.0458119661 -0.3113707  0.033816626
#>  [9,] 0.02603547  0.17014431  0.16974221  0.1195992322 -0.3161644  0.110227285
#> [10,] 0.15697446  0.18011359  0.18771176  0.2194333222 -0.2275454 -0.001443951
#>               V31         V32         V33          V34         V35          V36
#>  [1,]  0.14783862  0.03231598  0.05697218  0.068196193  0.01939276 -0.106772270
#>  [2,]  0.04808949  0.03060942 -0.03827129 -0.175036763 -0.11546337 -0.189223975
#>  [3,]  0.01929865 -0.07360247 -0.05995349 -0.123276569 -0.12197760 -0.098073040
#>  [4,] -0.09848789 -0.12377600 -0.01252721 -0.002646822 -0.12292718 -0.158597238
#>  [5,]  0.06906702 -0.09474703  0.08815948 -0.020362996  0.09844447  0.153841009
#>  [6,]  0.02496316  0.01006929 -0.08875249 -0.078127636 -0.09973256 -0.136242312
#>  [7,] -0.03179941  0.01100942 -0.03499278  0.078915208 -0.01965757 -0.011348884
#>  [8,]  0.02805657  0.11200617 -0.08141220  0.034389048  0.02550095 -0.047215871
#>  [9,] -0.05472612  0.01789856 -0.12694943 -0.111558499 -0.12533405 -0.129363994
#> [10,]  0.06194557 -0.07518583 -0.14606998 -0.063072905 -0.02695960 -0.007638206
#>               V37         V38         V39         V4           V40         V41
#>  [1,]  0.01934654 -0.09032297 -0.03976044 -0.2693918  0.0129198515 -0.03521766
#>  [2,] -0.17726244 -0.17706412 -0.17643248 -0.2674548 -0.0760957653 -0.08541589
#>  [3,] -0.16642372 -0.07299826 -0.10162916 -0.1832892 -0.0718110351 -0.10887070
#>  [4,] -0.03580716 -0.16509588 -0.14598440 -0.3340858 -0.0003518213 -0.13620950
#>  [5,]  0.11600265  0.11418636 -0.01959666 -0.2934523 -0.0174337894 -0.05101267
#>  [6,] -0.16178098 -0.05944298 -0.06156010 -0.2948849 -0.1047316399 -0.22909153
#>  [7,] -0.09449515 -0.08841993 -0.05364165 -0.1669233  0.0156119233  0.03472484
#>  [8,] -0.02020535  0.02558017 -0.08848718 -0.2162755 -0.1523877587 -0.12446033
#>  [9,]  0.02759893 -0.11713367 -0.10404787 -0.1913980 -0.1000985703 -0.07750385
#> [10,] -0.07847564 -0.11810468 -0.08211469 -0.2759595 -0.1900545791 -0.12079535
#>                V42          V43         V44         V45         V46         V47
#>  [1,] -0.090550910 -0.171130147 -0.05144864 -0.04357370 -0.05948685 -0.10425441
#>  [2,] -0.077549016 -0.055521030 -0.08566813 -0.16260939 -0.17252066 -0.17463722
#>  [3,] -0.137038306 -0.179893762 -0.16828076 -0.02681330 -0.15565574 -0.14422710
#>  [4,] -0.049859638 -0.155227198 -0.04229189 -0.19137729 -0.13808206 -0.24401275
#>  [5,] -0.087281212 -0.145494889 -0.03480942 -0.18777897 -0.14459266 -0.22800326
#>  [6,] -0.106608373 -0.276758779 -0.15348509 -0.09237624 -0.21083034 -0.19635923
#>  [7,]  0.008921682 -0.120296882 -0.08151919 -0.12796930 -0.06603589 -0.07341083
#>  [8,] -0.129497268  0.029519010 -0.10453513 -0.09983669 -0.18822593 -0.20229006
#>  [9,] -0.127809772 -0.004639106 -0.21426619 -0.12612081 -0.08337613 -0.18558623
#> [10,] -0.078606092 -0.034354762 -0.19916033 -0.18023289 -0.06487472 -0.15608396
#>              V48        V49         V5        V50        V51        V52
#>  [1,] -0.2696218 -0.2295746 -0.2710874 -0.4196368 -0.2552074 -0.4324861
#>  [2,] -0.2633723 -0.2711489 -0.1715867 -0.3040102 -0.3442195 -0.3101135
#>  [3,] -0.2233585 -0.1669617 -0.2816372 -0.1918502 -0.1971261 -0.3031398
#>  [4,] -0.1860663 -0.3702371 -0.2444684 -0.3531263 -0.4369617 -0.3394832
#>  [5,] -0.2722238 -0.2399492 -0.1475769 -0.2932119 -0.3258602 -0.3165602
#>  [6,] -0.1952859 -0.3450573 -0.3304729 -0.3734178 -0.4282353 -0.2853860
#>  [7,] -0.1559931 -0.2733609 -0.1747792 -0.1977643 -0.2852852 -0.2716623
#>  [8,] -0.2231171 -0.3320324 -0.2001315 -0.3647016 -0.3156448 -0.3675490
#>  [9,] -0.2978019 -0.2309723 -0.2630608 -0.2115684 -0.3089587 -0.3703553
#> [10,] -0.1472131 -0.2967326 -0.3228884 -0.2723601 -0.3346736 -0.2841296
#>              V53        V54        V55        V56        V57        V58
#>  [1,] -0.3326418 -0.4303410 -0.3361719 -0.3989270 -0.4278722 -0.4375290
#>  [2,] -0.3382993 -0.3181708 -0.3851246 -0.3652178 -0.3851429 -0.2963989
#>  [3,] -0.3183284 -0.2624755 -0.1873647 -0.2248500 -0.2855143 -0.3252712
#>  [4,] -0.3067009 -0.3524604 -0.3572832 -0.3730466 -0.3092060 -0.3776535
#>  [5,] -0.3010996 -0.2941506 -0.4247105 -0.4092810 -0.3032535 -0.3478652
#>  [6,] -0.4494234 -0.3879844 -0.3017196 -0.3524599 -0.4299602 -0.3378744
#>  [7,] -0.2101426 -0.1902949 -0.3869649 -0.2348342 -0.3780434 -0.2794087
#>  [8,] -0.3002353 -0.4196513 -0.3468394 -0.3238492 -0.2556363 -0.3068864
#>  [9,] -0.2878808 -0.2193105 -0.3254451 -0.3951349 -0.3785585 -0.2526175
#> [10,] -0.3645916 -0.2954450 -0.2595458 -0.2502074 -0.3963582 -0.3756542
#>              V59          V6        V60          V7          V8          V9
#>  [1,] -0.3473805 -0.13647744 -0.4495947 -0.24877095 -0.10720971 -0.11051032
#>  [2,] -0.2869945 -0.20212396 -0.3481372 -0.11685283 -0.10047508 -0.12624611
#>  [3,] -0.3096717 -0.08302964 -0.2677870 -0.06460681 -0.09261432 -0.14497203
#>  [4,] -0.4317201 -0.31533989 -0.4239842 -0.17237759 -0.22341055 -0.03072353
#>  [5,] -0.3453391 -0.15776238 -0.2500820 -0.21894481 -0.14364180 -0.19869441
#>  [6,] -0.3101414 -0.28926486 -0.3850923 -0.22485499 -0.22627761 -0.17694692
#>  [7,] -0.2884746 -0.18626220 -0.2527941 -0.08337543 -0.24770907 -0.20268142
#>  [8,] -0.3973070 -0.13246142 -0.3958420 -0.23377073 -0.19212633 -0.07023608
#>  [9,] -0.2870026 -0.16298992 -0.2135550 -0.11479019 -0.07774720 -0.13299951
#> [10,] -0.2491697 -0.14917047 -0.2311661 -0.23670595 -0.10931098 -0.13611611
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]       [,5]         [,6]
#> [1,] -0.09533736 -0.07333126 -0.11306311 -0.05598196 -0.2135624 0.0008483094
#> [2,]  0.03623372  0.09989724  0.07091964  0.09357774  0.2171714 0.0950202925
#>             [,7]        [,8]        [,9]      [,10]
#> [1,] 0.015003508 -0.03028068 -0.09885782 0.03585831
#> [2,] 0.007746395  0.03875195  0.11263461 0.00665130
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12          V13
#>  [1,]  2.347627e-05  2.133806e-04  2.263080e-04  0.0002692940 3.493739e-04
#>  [2,]  7.455212e-05  3.463303e-04  4.431878e-04  0.0004239690 5.979929e-04
#>  [3,]  8.140933e-05  3.518218e-04  4.925988e-04  0.0004621455 6.329642e-04
#>  [4,]  5.894147e-05  3.423731e-04  4.004426e-04  0.0004001340 5.531429e-04
#>  [5,]  1.445115e-04  3.509639e-04  6.896818e-04  0.0005831790 9.090644e-04
#>  [6,]  1.045215e-06 -2.422936e-05 -3.888174e-05 -0.0000373002 4.997789e-05
#>  [7,] -3.632821e-06  3.233258e-04  2.371673e-04  0.0003849023 4.276938e-04
#>  [8,]  9.485183e-06  2.461563e-04  2.024034e-04  0.0002859351 3.762977e-04
#>  [9,]  6.859312e-05  2.490919e-04  4.040789e-04  0.0003637139 5.317937e-04
#> [10,] -2.708160e-05  2.397702e-04  9.736841e-05  0.0002599923 2.793600e-04
#>                V14          V15          V16          V17           V18
#>  [1,] 0.0004489116 0.0004916501 0.0006524473 7.703020e-04  0.0008513026
#>  [2,] 0.0007085352 0.0006575794 0.0008081668 8.788309e-04  0.0008522189
#>  [3,] 0.0006869039 0.0005901117 0.0006759510 6.966390e-04  0.0006025559
#>  [4,] 0.0006980048 0.0007050459 0.0009002463 1.018218e-03  0.0010518216
#>  [5,] 0.0008368417 0.0004080987 0.0002737927 7.051662e-05 -0.0003730635
#>  [6,] 0.0001731333 0.0002264418 0.0003477489 4.216472e-04  0.0004410100
#>  [7,] 0.0006506992 0.0009143222 0.0013058132 1.641215e-03  0.0019081412
#>  [8,] 0.0005941232 0.0007769000 0.0011050336 1.361582e-03  0.0015474224
#>  [9,] 0.0005656171 0.0004449041 0.0004784486 4.624732e-04  0.0003192804
#> [10,] 0.0005750283 0.0009170927 0.0013755035 1.754158e-03  0.0021353908
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0008017333  3.497585e-05 0.0007370816 0.0007459775 0.0008987431
#>  [2,] 0.0010868762  1.213299e-04 0.0012923249 0.0013340784 0.0014315663
#>  [3,] 0.0009789988  1.444245e-04 0.0012449779 0.0012866738 0.0013668122
#>  [4,] 0.0011443230  9.302753e-05 0.0012402516 0.0012838292 0.0014455491
#>  [5,] 0.0007348850  3.004523e-04 0.0014401719 0.0014263937 0.0014274905
#>  [6,] 0.0003677700  1.016442e-05 0.0003024317 0.0003037623 0.0004199329
#>  [7,] 0.0013974357 -3.079982e-05 0.0009594821 0.0009649225 0.0013029632
#>  [8,] 0.0012286629  1.001825e-06 0.0009596965 0.0009764441 0.0012680610
#>  [9,] 0.0007382334  1.388473e-04 0.0009783856 0.0009822399 0.0010694704
#> [10,] 0.0014581814 -7.351181e-05 0.0008995791 0.0009052773 0.0013041398
#>                V23          V24           V25           V26           V27
#>  [1,] 0.0010847842 1.114649e-03  0.0011410386  1.165052e-03  1.134537e-03
#>  [2,] 0.0013369950 1.241301e-03  0.0011480436  1.020189e-03  8.649906e-04
#>  [3,] 0.0011369687 9.728486e-04  0.0008020812  6.094262e-04  4.635885e-04
#>  [4,] 0.0015161300 1.491825e-03  0.0014578345  1.399625e-03  1.292646e-03
#>  [5,] 0.0005443546 3.387935e-05 -0.0004356206 -1.025886e-03 -1.389724e-03
#>  [6,] 0.0005725301 6.385586e-04  0.0007057418  6.844231e-04  6.188644e-04
#>  [7,] 0.0020530725 2.281574e-03  0.0024375418  2.664161e-03  2.819487e-03
#>  [8,] 0.0017717575 1.911704e-03  0.0020243482  2.132504e-03  2.156993e-03
#>  [9,] 0.0007615025 5.467056e-04  0.0003537355  7.553903e-05 -9.123868e-05
#> [10,] 0.0022473915 2.560313e-03  0.0028051319  3.117996e-03  3.287257e-03
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0010821442  1.027847e-03  4.596006e-05  0.0009028354  9.073413e-04
#>  [2,]  0.0007322503  5.100811e-04  1.259370e-04  0.0005144257  3.917299e-04
#>  [3,]  0.0003394011  5.564381e-05  1.380708e-04  0.0001008999 -6.466833e-05
#>  [4,]  0.0011795524  9.980945e-04  1.032225e-04  0.0009338514  8.586393e-04
#>  [5,] -0.0015945334 -2.135993e-03  2.567740e-04 -0.0018527023 -2.284962e-03
#>  [6,]  0.0005393950  4.975617e-04  1.397780e-05  0.0004561258  4.644612e-04
#>  [7,]  0.0028389400  2.894994e-03 -1.378425e-06  0.0025299741  2.674735e-03
#>  [8,]  0.0020854700  2.072658e-03  2.497806e-05  0.0018376043  1.918676e-03
#>  [9,] -0.0002178163 -4.858397e-04  1.249878e-04 -0.0004129119 -5.779680e-04
#> [10,]  0.0032764104  3.429027e-03 -3.407729e-05  0.0030148844  3.229191e-03
#>                 V32           V33           V34           V35           V36
#>  [1,]  8.241983e-04  0.0008604891  7.750771e-04  8.018590e-04  0.0009091675
#>  [2,]  6.383254e-04  0.0005790121  3.842280e-04  6.332619e-05 -0.0001923848
#>  [3,]  3.309555e-04  0.0002042064 -3.761994e-05 -5.234079e-04 -0.0009262935
#>  [4,]  9.417392e-04  0.0009229182  7.596343e-04  6.071277e-04  0.0005228339
#>  [5,] -9.174378e-04 -0.0012240966 -1.712168e-03 -3.102537e-03 -0.0044418731
#>  [6,]  4.692662e-04  0.0005430498  4.921055e-04  5.034440e-04  0.0005371374
#>  [7,]  2.010105e-03  0.0021588868  2.194066e-03  2.747349e-03  0.0033969038
#>  [8,]  1.541038e-03  0.0016903428  1.653124e-03  1.955698e-03  0.0023406821
#>  [9,] -5.481749e-05 -0.0001981095 -4.546701e-04 -1.015704e-03 -0.0015087826
#> [10,]  2.383190e-03  0.0026693663  2.760964e-03  3.510589e-03  0.0043459435
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0008347564  6.596005e-04  5.743815e-04  5.879272e-05  0.0005845656
#>  [2,] -0.0001906026  2.669267e-04  5.591058e-04  1.657590e-04  0.0002341235
#>  [3,] -0.0008253258 -6.405129e-05  4.344237e-04  1.780145e-04  0.0000163547
#>  [4,]  0.0004588143  5.845631e-04  6.808288e-04  1.354789e-04  0.0004956941
#>  [5,] -0.0040962602 -1.597802e-03 -1.135711e-05  3.204876e-04 -0.0011814754
#>  [6,]  0.0004329075  3.080618e-04  2.456478e-04  1.295862e-05  0.0002247074
#>  [7,]  0.0031109962  1.931339e-03  1.215681e-03 -1.211488e-05  0.0016308954
#>  [8,]  0.0021028622  1.358970e-03  9.198144e-04  3.102018e-05  0.0011502606
#>  [9,] -0.0013547313 -4.042388e-04  2.106531e-04  1.569925e-04 -0.0002589133
#> [10,]  0.0039199015  2.290314e-03  1.296275e-03 -4.791352e-05  0.0018979842
#>                 V41           V42          V43           V44           V45
#>  [1,]  0.0004939947  3.945024e-04 3.165254e-04  2.178266e-04  1.326295e-04
#>  [2,]  0.0002676250  2.642863e-04 4.388033e-04  5.146022e-04  6.389767e-04
#>  [3,]  0.0001050833  1.443580e-04 4.157767e-04  5.799202e-04  8.055408e-04
#>  [4,]  0.0004503829  3.833730e-04 4.437916e-04  4.314657e-04  4.549102e-04
#>  [5,] -0.0007303394 -4.152110e-04 4.785393e-04  1.074241e-03  1.848924e-03
#>  [6,]  0.0001668495  8.203574e-05 8.595186e-05  4.429889e-05  2.111412e-06
#>  [7,]  0.0012584192  9.194669e-04 4.748337e-04  8.088819e-05 -3.598530e-04
#>  [8,]  0.0008999737  6.544254e-04 3.997316e-04  1.380024e-04 -1.442614e-04
#>  [9,] -0.0001279771 -6.010261e-05 2.962307e-04  5.238574e-04  7.974020e-04
#> [10,]  0.0014130540  9.946947e-04 4.279379e-04 -7.625652e-05 -6.648524e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.160198e-04  9.868525e-05  6.033712e-05  3.755986e-05  9.385072e-05
#>  [2,]  5.717821e-04  3.497601e-04  2.389088e-04  1.424643e-04  2.325366e-04
#>  [3,]  7.282293e-04  4.368853e-04  2.983765e-04  1.762594e-04  2.581088e-04
#>  [4,]  4.026947e-04  2.563212e-04  1.710150e-04  1.007476e-04  1.914440e-04
#>  [5,]  1.690187e-03  9.231881e-04  6.079579e-04  3.681033e-04  4.957572e-04
#>  [6,] -1.274412e-05 -2.530057e-05 -3.207075e-05 -2.084664e-05  2.603004e-05
#>  [7,] -3.718382e-04 -1.433615e-04 -1.135453e-04 -6.758525e-05  1.256193e-05
#>  [8,] -1.556829e-04 -4.567247e-05 -4.631783e-05 -2.788175e-05  6.492919e-05
#>  [9,]  7.198520e-04  4.138315e-04  2.705439e-04  1.619805e-04  2.421520e-04
#> [10,] -6.492334e-04 -3.062418e-04 -2.345727e-04 -1.420718e-04 -3.781185e-05
#>                V50           V51          V52          V53          V54
#>  [1,] 2.769994e-05  1.827272e-05 1.633488e-05 1.760936e-05 1.479408e-05
#>  [2,] 5.904949e-05  4.924710e-05 3.559643e-05 2.748353e-05 2.311515e-05
#>  [3,] 6.261798e-05  5.726212e-05 3.707508e-05 2.654310e-05 2.221001e-05
#>  [4,] 5.184682e-05  3.998640e-05 3.167225e-05 2.708264e-05 2.314684e-05
#>  [5,] 1.013947e-04  1.113479e-04 5.877463e-05 3.601644e-05 2.578900e-05
#>  [6,] 1.030689e-05  4.913233e-06 5.116018e-06 1.098302e-05 7.290258e-06
#>  [7,] 1.718585e-05 -4.805889e-06 1.006467e-05 2.371557e-05 2.123643e-05
#>  [8,] 2.479720e-05  7.877902e-06 1.472370e-05 2.334747e-05 1.946673e-05
#>  [9,] 5.236764e-05  5.310324e-05 3.088430e-05 2.112183e-05 1.593757e-05
#> [10,] 3.964359e-06 -1.802228e-05 4.851357e-06 2.294865e-05 1.944062e-05
#>                V55           V56          V57           V58           V59
#>  [1,] 1.402541e-05  9.643852e-06 1.188307e-05  8.361168e-06  1.155258e-05
#>  [2,] 2.733189e-05  2.635788e-05 1.565049e-05  2.386901e-05  2.934196e-05
#>  [3,] 2.742740e-05  2.929478e-05 1.272113e-05  2.728913e-05  3.085813e-05
#>  [4,] 2.471021e-05  2.157001e-05 1.775552e-05  1.924300e-05  2.485246e-05
#>  [5,] 4.615962e-05  5.467709e-05 6.039843e-06  5.451027e-05  5.688843e-05
#>  [6,] 7.416132e-06  4.129168e-06 6.437819e-06  2.467793e-06  7.296619e-06
#>  [7,] 1.151301e-05  3.803277e-08 2.307247e-05 -1.335916e-06  2.480615e-06
#>  [8,] 1.422978e-05  5.247172e-06 1.965431e-05  3.018547e-06  9.226524e-06
#>  [9,] 2.327135e-05  2.591188e-05 6.806734e-06  2.518713e-05  2.738136e-05
#> [10,] 5.429704e-06 -8.478134e-06 2.427234e-05 -1.031866e-05 -3.448684e-06
#>                 V6          V60           V7            V8            V9
#>  [1,] 1.447224e-04 1.197383e-05 2.039087e-04  1.932669e-04  1.419946e-04
#>  [2,] 2.015827e-04 2.239315e-05 2.176167e-04  1.843897e-04  3.219044e-04
#>  [3,] 1.859542e-04 2.192292e-05 1.665140e-04  1.300888e-04  3.513669e-04
#>  [4,] 2.093731e-04 2.106890e-05 2.598068e-04  2.361638e-04  2.896888e-04
#>  [5,] 1.934276e-04 3.251598e-05 5.623955e-06 -1.174767e-04  4.872967e-04
#>  [6,] 4.545833e-05 6.558059e-06 8.592087e-05  3.755155e-05 -2.485013e-05
#>  [7,] 2.484302e-04 1.096236e-05 4.308158e-04  4.494369e-04  1.526662e-04
#>  [8,] 2.068109e-04 1.399698e-05 3.459414e-04  3.290201e-04  1.216618e-04
#>  [9,] 1.464502e-04 1.782375e-05 9.823537e-05  4.646823e-05  2.812635e-04
#> [10,] 2.431277e-04 9.660492e-06 4.694144e-04  4.800222e-04  4.329976e-05
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.07232677  0.05257715  0.04510251  0.05647973  0.07073701  0.03824124
#> [2,] -0.07232677 -0.05257715 -0.04510251 -0.05647973 -0.07073701 -0.03824124
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.06479276  0.06561743  0.04783195  0.06259125
#> [2,] -0.06479276 -0.06561743 -0.04783195 -0.06259125
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6262961 0.3085509 0.3328567 0.6334851 0.5517768 0.3517599 0.4106579
#>  [8] 0.6355470 0.2691431 0.4770977
#> 
#> $B[[2]]
#> [1] -0.04794230 -0.03958241
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  1.750638e-03  1.422143e-03  7.940978e-04  1.979251e-03 -1.801380e-03
#>  [6]  8.924233e-04  4.132348e-03  3.209312e-03 -1.260474e-05  4.771279e-03
#> 
#> $vB[[2]]
#> [1]  0.08208811 -0.08208811
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#>  [2,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#>  [3,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#>  [4,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#>  [5,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#>  [6,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#>  [7,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#>  [8,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#>  [9,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [10,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [11,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [12,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [13,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [14,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [15,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [16,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [17,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [18,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [19,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [20,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [21,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [22,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [23,] 0.0135 0.1908 0.1576 0.1112 0.1197 0.1174 0.1415 0.2215 0.2658 0.2713
#> [24,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [25,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [26,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [27,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [28,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [29,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [30,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [31,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [32,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [33,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [34,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [35,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [36,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [37,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [38,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [39,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [40,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#>  [2,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#>  [3,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#>  [4,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#>  [5,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#>  [6,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#>  [7,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#>  [8,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#>  [9,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [10,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [11,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [12,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [13,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [14,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [15,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [16,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [17,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [18,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [19,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [20,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [21,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [22,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [23,] 0.3862 0.0045 0.5717 0.6797 0.8747 1.0000 0.8948 0.8420 0.9174 0.9307
#> [24,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [25,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [26,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [27,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [28,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [29,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [30,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [31,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [32,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [33,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [34,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [35,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [36,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [37,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [38,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [39,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [40,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#>  [2,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#>  [3,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#>  [4,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#>  [5,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#>  [6,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#>  [7,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#>  [8,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#>  [9,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [10,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [11,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [12,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [13,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [14,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [15,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [16,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [17,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [18,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [19,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [20,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [21,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [22,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [23,] 0.9050 0.8228 0.0051 0.6986 0.5831 0.4924 0.4563 0.5159 0.5670 0.5284
#> [24,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [25,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [26,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [27,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [28,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [29,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [30,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [31,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [32,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [33,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [34,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [35,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [36,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [37,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [38,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [39,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [40,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#>  [2,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#>  [3,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#>  [4,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#>  [5,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#>  [6,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#>  [7,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#>  [8,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#>  [9,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [10,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [11,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [12,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [13,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [14,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [15,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [16,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [17,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [18,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [19,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [20,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [21,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [22,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [23,] 0.5144 0.3742 0.2282 0.0289 0.1193 0.1088 0.0431 0.1070 0.0583 0.0046
#> [24,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [25,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [26,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [27,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [28,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [29,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [30,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [31,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [32,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [33,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [34,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [35,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [36,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [37,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [38,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [39,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [40,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#>  [2,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#>  [3,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#>  [4,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#>  [5,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#>  [6,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#>  [7,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#>  [8,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#>  [9,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [10,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [11,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [12,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [13,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [14,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [15,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [16,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [17,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [18,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [19,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [20,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [21,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [22,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [23,] 0.0473 0.0408 0.0290 0.0192 0.0561 0.0094 0.0025 0.0037 0.0084 0.0102
#> [24,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [25,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [26,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [27,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [28,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [29,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [30,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [31,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [32,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [33,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [34,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [35,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [36,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [37,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [38,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [39,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [40,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#>  [2,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#>  [3,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#>  [4,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#>  [5,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#>  [6,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#>  [7,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#>  [8,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#>  [9,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [10,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [11,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [12,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [13,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [14,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [15,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [16,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [17,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [18,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [19,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [20,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [21,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [22,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [23,] 0.0096 0.0024 0.0037 0.0028 0.0030 0.0929 0.0030 0.1031 0.0883 0.1596
#> [24,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [25,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [26,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [27,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [28,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [29,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [30,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [31,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [32,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [33,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [34,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [35,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [36,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [37,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [38,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [39,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [40,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9269665 0.8244426 0.6782984 0.8068409 0.6327442 0.3582959 0.8611339
#>  [2,] 0.9100838 0.7882834 0.6429317 0.7665604 0.6734842 0.3942058 0.8358591
#>  [3,] 0.8704777 0.6664198 0.6198847 0.7204907 0.6021925 0.2687191 0.7683095
#>  [4,] 0.7860470 0.5366086 0.6026594 0.6563321 0.4801610 0.1639554 0.6686274
#>  [5,] 0.8302662 0.5607082 0.4369112 0.5536678 0.5577515 0.1752613 0.7319572
#>  [6,] 0.9046994 0.7210518 0.6000006 0.7278900 0.5592425 0.2104712 0.8123717
#>  [7,] 0.8481930 0.5470420 0.5194568 0.6131941 0.4632034 0.1193973 0.6869940
#>  [8,] 0.8158083 0.5739018 0.4575142 0.5586220 0.5278181 0.2231731 0.6969618
#>  [9,] 0.7861174 0.4485454 0.3706959 0.5019020 0.6881830 0.2889254 0.6661224
#> [10,] 0.8878358 0.7217070 0.5831288 0.7223762 0.6383267 0.2772514 0.8237446
#> [11,] 0.8854090 0.6493908 0.5952103 0.7237297 0.5612285 0.1447900 0.8032379
#> [12,] 0.9019670 0.7578765 0.5991457 0.7420718 0.6578706 0.3793244 0.8202127
#> [13,] 0.8635117 0.6156757 0.5294607 0.6241849 0.4493205 0.1391211 0.7530966
#> [14,] 0.8918836 0.7312194 0.6864091 0.7621182 0.4681924 0.1403331 0.7929652
#> [15,] 0.8805078 0.7228220 0.5703496 0.7027726 0.6296480 0.3711308 0.7944669
#> [16,] 0.8856346 0.7297606 0.6846438 0.7631483 0.5531247 0.2334169 0.7945907
#> [17,] 0.8847232 0.6757903 0.6226697 0.7241303 0.5697621 0.1806841 0.8053364
#> [18,] 0.7490936 0.5254167 0.5995114 0.6531363 0.4488055 0.1622065 0.6338264
#> [19,] 0.8429958 0.5834728 0.6070132 0.6791850 0.5394202 0.1594389 0.6911305
#> [20,] 0.9296122 0.8289677 0.7391030 0.8288043 0.6212336 0.3145031 0.8684368
#> [21,] 0.8743015 0.6687693 0.5216284 0.6649789 0.5508538 0.2102788 0.7683021
#> [22,] 0.7667088 0.3866667 0.3783346 0.4785414 0.5907058 0.1539371 0.6317821
#> [23,] 0.9294260 0.7959474 0.6388483 0.7776798 0.7118869 0.3569809 0.8586537
#> [24,] 0.8999108 0.7240533 0.6341049 0.7500557 0.6071934 0.2648733 0.8250128
#> [25,] 0.7631198 0.4500372 0.4101366 0.5097668 0.5790688 0.1894465 0.6613211
#> [26,] 0.8334921 0.5442085 0.4207947 0.5801702 0.6224508 0.2038961 0.7327678
#> [27,] 0.9142676 0.7899786 0.7067567 0.8001743 0.5610306 0.2318523 0.8288492
#> [28,] 0.8473031 0.7048844 0.6277038 0.7137336 0.5927293 0.2993245 0.7537393
#> [29,] 0.8745406 0.7197259 0.6147116 0.7338305 0.6241557 0.3504440 0.7850789
#> [30,] 0.8484480 0.7186760 0.5977106 0.7154742 0.6767677 0.4572871 0.7763935
#> [31,] 0.8812301 0.6964381 0.5623066 0.6860406 0.6215992 0.3097679 0.7848206
#> [32,] 0.8708569 0.7138339 0.5663668 0.7078148 0.6438396 0.3938044 0.7850512
#> [33,] 0.8744855 0.6718746 0.6127916 0.7225825 0.5864043 0.1986065 0.7924659
#> [34,] 0.9244798 0.7649697 0.6878764 0.7857422 0.5532042 0.1787593 0.8397560
#> [35,] 0.9099579 0.7242945 0.5274591 0.6927708 0.7028271 0.3774925 0.8345361
#> [36,] 0.8764697 0.6839387 0.6507013 0.7485847 0.5917081 0.2342488 0.7976181
#> [37,] 0.7517673 0.4005824 0.3524442 0.4712756 0.6727698 0.2659384 0.6401586
#> [38,] 0.8662747 0.6388069 0.5981472 0.6712490 0.5324688 0.1487370 0.7492634
#> [39,] 0.8917303 0.6640215 0.6010689 0.6979737 0.5846507 0.1846166 0.7768697
#> [40,] 0.8751169 0.7459969 0.6284469 0.7640241 0.6095357 0.2972773 0.8146835
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8221859 0.6032281 0.8159021
#>  [2,] 0.8069926 0.5981145 0.7953921
#>  [3,] 0.7631826 0.5137718 0.7133333
#>  [4,] 0.6776752 0.3839428 0.6094809
#>  [5,] 0.6681081 0.4115599 0.6260559
#>  [6,] 0.7603680 0.4788661 0.7166540
#>  [7,] 0.6752266 0.3735722 0.6025240
#>  [8,] 0.6546798 0.4064526 0.5966903
#>  [9,] 0.6519789 0.4185320 0.5701824
#> [10,] 0.7805703 0.5014577 0.7271760
#> [11,] 0.7508509 0.3940738 0.6803876
#> [12,] 0.7826915 0.5704310 0.7586477
#> [13,] 0.6936456 0.3600119 0.6012989
#> [14,] 0.7506292 0.4404099 0.7233054
#> [15,] 0.7514937 0.5470606 0.7385589
#> [16,] 0.7708980 0.5081429 0.7299208
#> [17,] 0.7700138 0.4284797 0.6901757
#> [18,] 0.6506333 0.3766827 0.5930959
#> [19,] 0.6909062 0.4246492 0.6361919
#> [20,] 0.8305860 0.5854331 0.8056206
#> [21,] 0.7205443 0.4508730 0.6710711
#> [22,] 0.6180342 0.3267902 0.5018427
#> [23,] 0.8321590 0.5943057 0.8074355
#> [24,] 0.7872995 0.4746795 0.7216175
#> [25,] 0.6214285 0.3558919 0.5320996
#> [26,] 0.6431534 0.4190601 0.6000163
#> [27,] 0.7924224 0.5659342 0.7846607
#> [28,] 0.7496270 0.5560024 0.7556901
#> [29,] 0.7657935 0.5441744 0.7308287
#> [30,] 0.7545598 0.5656902 0.7461351
#> [31,] 0.7533303 0.5230956 0.7042917
#> [32,] 0.7553700 0.5358961 0.7188955
#> [33,] 0.7621663 0.4617506 0.6977907
#> [34,] 0.7958234 0.4931744 0.7763913
#> [35,] 0.7900651 0.5215864 0.7344112
#> [36,] 0.7682232 0.4733876 0.7213173
#> [37,] 0.6215329 0.3858716 0.5214023
#> [38,] 0.7080270 0.4138826 0.6910514
#> [39,] 0.7313772 0.4552352 0.7170876
#> [40,] 0.7710399 0.5517124 0.7394154
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5135947 0.4864053
#>  [2,] 0.5105049 0.4894951
#>  [3,] 0.5154012 0.4845988
#>  [4,] 0.5226137 0.4773863
#>  [5,] 0.5233687 0.4766313
#>  [6,] 0.5209405 0.4790595
#>  [7,] 0.5270192 0.4729808
#>  [8,] 0.5221323 0.4778677
#>  [9,] 0.5123774 0.4876226
#> [10,] 0.5157681 0.4842319
#> [11,] 0.5229853 0.4770147
#> [12,] 0.5119109 0.4880891
#> [13,] 0.5293686 0.4706314
#> [14,] 0.5258252 0.4741748
#> [15,] 0.5140831 0.4859169
#> [16,] 0.5179881 0.4820119
#> [17,] 0.5208659 0.4791341
#> [18,] 0.5230274 0.4769726
#> [19,] 0.5182610 0.4817390
#> [20,] 0.5134249 0.4865751
#> [21,] 0.5216743 0.4783257
#> [22,] 0.5193567 0.4806433
#> [23,] 0.5095660 0.4904340
#> [24,] 0.5173326 0.4826674
#> [25,] 0.5193083 0.4806917
#> [26,] 0.5174175 0.4825825
#> [27,] 0.5175815 0.4824185
#> [28,] 0.5148015 0.4851985
#> [29,] 0.5129680 0.4870320
#> [30,] 0.5086878 0.4913122
#> [31,] 0.5147989 0.4852011
#> [32,] 0.5124289 0.4875711
#> [33,] 0.5186337 0.4813663
#> [34,] 0.5216235 0.4783765
#> [35,] 0.5124807 0.4875193
#> [36,] 0.5176099 0.4823901
#> [37,] 0.5127068 0.4872932
#> [38,] 0.5229022 0.4770978
#> [39,] 0.5194324 0.4805676
#> [40,] 0.5144233 0.4855767
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]       [,2]       [,3]         [,4]       [,5]       [,6]
#>  [1,] 2.541000  1.5467417  0.7459628  1.429611977  0.5440076 -0.5827682
#>  [2,] 2.314659  1.3146091  0.5881114  1.188990072  0.7239864 -0.4296673
#>  [3,] 1.905189  0.6920367  0.4890587  0.946896990  0.4146091 -1.0011313
#>  [4,] 1.301260  0.1466970  0.4165582  0.646991262 -0.0793977 -1.6290876
#>  [5,] 1.587515  0.2440366 -0.2537074  0.215501224  0.2320416 -1.5487886
#>  [6,] 2.250567  0.9496851  0.4054674  0.983943767  0.2380885 -1.3220874
#>  [7,] 1.720498  0.1887260  0.0778664  0.460758469 -0.1474531 -1.9981501
#>  [8,] 1.488202  0.2977884 -0.1703542  0.235571554  0.1113873 -1.2472701
#>  [9,] 1.301679 -0.2065495 -0.5292327  0.007608195  0.7916384 -0.9006090
#> [10,] 2.068823  0.9529449  0.3356307  0.956279193  0.5681090 -0.9581369
#> [11,] 2.044680  0.6163625  0.3855473  0.963038010  0.2461494 -1.7760629
#> [12,] 2.219274  1.1410728  0.4019066  1.056764601  0.6538194 -0.4924169
#> [13,] 1.844768  0.4712335  0.1179796  0.507349394 -0.2034165 -1.8226093
#> [14,] 2.110127  1.0008181  0.7833845  1.164327477 -0.1274024 -1.8125265
#> [15,] 1.997248  0.9585032  0.2832775  0.860536064  0.5307069 -0.5273688
#> [16,] 2.046906  0.9934083  0.7751960  1.170017988  0.2133037 -1.1891170
#> [17,] 2.037938  0.7344924  0.5008951  0.965042588  0.2808808 -1.5117193
#> [18,] 1.093784  0.1017543  0.4034298  0.632853938 -0.2054980 -1.6419018
#> [19,] 1.680689  0.3370459  0.4347745  0.750029086  0.1580087 -1.6624089
#> [20,] 2.580747  1.5783286  1.0413113  1.577176529  0.4947877 -0.7791500
#> [21,] 1.939539  0.7026239  0.0865675  0.685561845  0.2041211 -1.3232458
#> [22,] 1.189819 -0.4613455 -0.4966231 -0.085887168  0.3668838 -1.7040493
#> [23,] 2.577906  1.3611550  0.5703690  1.252196354  0.9045659 -0.5884921
#> [24,] 2.196234  0.9646573  0.5498679  1.098909211  0.4355299 -1.0207914
#> [25,] 1.169861 -0.2005202 -0.3634007  0.039072252  0.3189520 -1.4536110
#> [26,] 1.610582  0.1772971 -0.3195121  0.323471982  0.4999636 -1.3621190
#> [27,] 2.366892  1.3247962  0.8796838  1.387384274  0.2453457 -1.1978816
#> [28,] 1.713604  0.8706668  0.5223785  0.913586725  0.3752598 -0.8505166
#> [29,] 1.941717  0.9431024  0.4671610  1.014144920  0.5072250 -0.6170880
#> [30,] 1.722480  0.9379034  0.3959348  0.922121841  0.7389570 -0.1712692
#> [31,] 2.004131  0.8303936  0.2505284  0.781672997  0.4963417 -0.8012048
#> [32,] 1.908557  0.9140779  0.2670429  0.884794399  0.5920686 -0.4313484
#> [33,] 1.941214  0.7166759  0.4590617  0.957307944  0.3491206 -1.3950267
#> [34,] 2.504830  1.1801219  0.7902098  1.299448698  0.2136256 -1.5247761
#> [35,] 2.313121  0.9658650  0.1099470  0.813105043  0.8607966 -0.5002048
#> [36,] 1.959416  0.7719322  0.6221232  1.091078066  0.3710310 -1.1844737
#> [37,] 1.108060 -0.4030389 -0.6083124 -0.115024057  0.7207395 -1.0153283
#> [38,] 1.868414  0.5701893  0.3977509  0.713839676  0.1300583 -1.7445412
#> [39,] 2.108539  0.6812675  0.4099209  0.837667105  0.3418946 -1.4853774
#> [40,] 1.946980  1.0773752  0.5255598  1.174869581  0.4453609 -0.8602972
#>            [,7]      [,8]         [,9]       [,10]
#>  [1,] 1.8247395 1.5312280  0.418933778 1.488826569
#>  [2,] 1.6277351 1.4305861  0.397614829 1.357739912
#>  [3,] 1.1987903 1.1702080  0.055101302 0.911628369
#>  [4,] 0.7019835 0.7431082 -0.472846160 0.445130584
#>  [5,] 1.0045755 0.6996406 -0.357520701 0.515333219
#>  [6,] 1.4654953 1.1546980 -0.084585937 0.927924152
#>  [7,] 0.7861031 0.7319205 -0.516922382 0.415993099
#>  [8,] 0.8328716 0.6396742 -0.378649926 0.391693476
#>  [9,] 0.6906988 0.6277489 -0.328802713 0.282595346
#> [10,] 1.5419265 1.2689929  0.005830783 0.980341567
#> [11,] 1.4066556 1.1031557 -0.430219954 0.755553655
#> [12,] 1.5177892 1.2814209  0.283609746 1.145279747
#> [13,] 1.1151962 0.8172187 -0.575312574 0.410880162
#> [14,] 1.3428926 1.1019708 -0.239498663 0.960916971
#> [15,] 1.3520645 1.1065947  0.188801118 1.038491721
#> [16,] 1.3528224 1.2133885  0.032574413 0.994220741
#> [17,] 1.4199869 1.2083891 -0.288056754 0.800940732
#> [18,] 0.5486675 0.6218242 -0.503652425 0.376778768
#> [19,] 0.8054097 0.8043595 -0.303716450 0.558873584
#> [20,] 1.8872076 1.5897863  0.345117433 1.421801257
#> [21,] 1.1987488 0.9471630 -0.197144001 0.713033446
#> [22,] 0.5398695 0.4812128 -0.722738789 0.007370997
#> [23,] 1.8041528 1.6010068  0.381793795 1.433431901
#> [24,] 1.5506861 1.3087238 -0.101368573 0.952499006
#> [25,] 0.6691873 0.4956161 -0.593239593 0.128575153
#> [26,] 1.0087109 0.5890771 -0.326632865 0.405532953
#> [27,] 1.5774932 1.3395896  0.265281778 1.293036443
#> [28,] 1.1186558 1.0966238  0.224953419 1.129193732
#> [29,] 1.2955134 1.1847093  0.177159601 0.998831332
#> [30,] 1.2447718 1.1230812  0.264288544 1.078104227
#> [31,] 1.2939830 1.1164537  0.092448100 0.867819314
#> [32,] 1.2953493 1.1274610  0.143831896 0.938989605
#> [33,] 1.3398539 1.1645932 -0.153297248 0.836799256
#> [34,] 1.6564141 1.3603919 -0.027303972 1.244758909
#> [35,] 1.6181232 1.3253178  0.086399291 1.017119792
#> [36,] 1.3714734 1.1983055 -0.106550309 0.951005303
#> [37,] 0.5760525 0.4960596 -0.464699275 0.085661691
#> [38,] 1.0946874 0.8858208 -0.347937700 0.805039480
#> [39,] 1.2475168 1.0016213 -0.179539934 0.930060905
#> [40,] 1.4807347 1.2141923  0.207591864 1.042932051
#> 
#> $pre[[3]]
#>              [,1]        [,2]
#>  [1,] 0.020956556 -0.03343547
#>  [2,] 0.015751989 -0.02627393
#>  [3,] 0.015299608 -0.04632452
#>  [4,] 0.019833865 -0.07068277
#>  [5,] 0.023827516 -0.06971556
#>  [6,] 0.024115350 -0.05969568
#>  [7,] 0.025723899 -0.08245848
#>  [8,] 0.022739319 -0.06584775
#>  [9,] 0.005779849 -0.04373982
#> [10,] 0.018453377 -0.04463992
#> [11,] 0.023209920 -0.06879625
#> [12,] 0.016869528 -0.03078311
#> [13,] 0.032523314 -0.08508660
#> [14,] 0.029330879 -0.07406206
#> [15,] 0.019796825 -0.03655056
#> [16,] 0.019174643 -0.05280873
#> [17,] 0.021048764 -0.06246331
#> [18,] 0.020563601 -0.07161140
#> [19,] 0.011441850 -0.06163477
#> [20,] 0.017387925 -0.03632458
#> [21,] 0.024402337 -0.06234931
#> [22,] 0.009639001 -0.06782662
#> [23,] 0.012579155 -0.02568941
#> [24,] 0.019797144 -0.04956110
#> [25,] 0.013281626 -0.06398980
#> [26,] 0.013319393 -0.05637873
#> [27,] 0.020525884 -0.04982902
#> [28,] 0.017421256 -0.04180217
#> [29,] 0.016117197 -0.03576626
#> [30,] 0.013866783 -0.02088811
#> [31,] 0.016761379 -0.04245138
#> [32,] 0.017357455 -0.03236831
#> [33,] 0.018107509 -0.05646184
#> [34,] 0.025017306 -0.06153049
#> [35,] 0.017247193 -0.03268585
#> [36,] 0.018231326 -0.05223744
#> [37,] 0.003810158 -0.04702785
#> [38,] 0.022124926 -0.06954820
#> [39,] 0.018261438 -0.05950739
#> [40,] 0.017658562 -0.04005046
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4864053 -0.4864053
#>  [2,] -0.5105049  0.5105049
#>  [3,] -0.5154012  0.5154012
#>  [4,]  0.4773863 -0.4773863
#>  [5,]  0.4766313 -0.4766313
#>  [6,]  0.4790595 -0.4790595
#>  [7,] -0.5270192  0.5270192
#>  [8,]  0.4778677 -0.4778677
#>  [9,] -0.5123774  0.5123774
#> [10,] -0.5157681  0.5157681
#> [11,] -0.5229853  0.5229853
#> [12,]  0.4880891 -0.4880891
#> [13,]  0.4706314 -0.4706314
#> [14,]  0.4741748 -0.4741748
#> [15,]  0.4859169 -0.4859169
#> [16,]  0.4820119 -0.4820119
#> [17,]  0.4791341 -0.4791341
#> [18,]  0.4769726 -0.4769726
#> [19,] -0.5182610  0.5182610
#> [20,] -0.5134249  0.5134249
#> [21,]  0.4783257 -0.4783257
#> [22,] -0.5193567  0.5193567
#> [23,] -0.5095660  0.5095660
#> [24,]  0.4826674 -0.4826674
#> [25,]  0.4806917 -0.4806917
#> [26,] -0.5174175  0.5174175
#> [27,]  0.4824185 -0.4824185
#> [28,] -0.5148015  0.5148015
#> [29,] -0.5129680  0.5129680
#> [30,] -0.5086878  0.5086878
#> [31,]  0.4852011 -0.4852011
#> [32,] -0.5124289  0.5124289
#> [33,]  0.4813663 -0.4813663
#> [34,]  0.4783765 -0.4783765
#> [35,] -0.5124807  0.5124807
#> [36,] -0.5176099  0.5176099
#> [37,] -0.5127068  0.5127068
#> [38,] -0.5229022  0.5229022
#> [39,] -0.5194324  0.5194324
#> [40,] -0.5144233  0.5144233
#> 
#> $L
#> [1] 0.7054599 0.7328864 0.6975792 0.7058555 0.7844506 0.6915091
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5217391