Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.3391082 -0.15892605 -0.02006695 -0.073100489 -0.04965016 -0.057843722
#>  [2,] -0.2081668 -0.14833617 -0.01861613 -0.172341773 -0.09420691 -0.212591926
#>  [3,] -0.3557483 -0.15026201 -0.09790937  0.007999057 -0.05743454 -0.128772348
#>  [4,] -0.3240607 -0.06712128 -0.08427500 -0.119281095 -0.16718678 -0.002258784
#>  [5,] -0.2552969 -0.09598568 -0.12482469 -0.176852929 -0.02501008  0.004904119
#>  [6,] -0.2005322 -0.01644657 -0.14287052 -0.120209435  0.00674683 -0.091266123
#>  [7,] -0.2637195 -0.15050167 -0.15489098 -0.110311942 -0.09890093 -0.022313744
#>  [8,] -0.3531902 -0.18695801 -0.06518218 -0.176932174 -0.11250839 -0.057188659
#>  [9,] -0.2579140 -0.09710780 -0.05770875 -0.024943585 -0.13667446 -0.076221680
#> [10,] -0.3736541 -0.15959419 -0.13890369 -0.144329458 -0.03599057  0.033270137
#>               V15         V16         V17          V18         V19         V2
#>  [1,]  0.04101134  0.04130155  0.01755577 -0.016577319 -0.04035964 -0.2969111
#>  [2,] -0.01437017 -0.05271413 -0.10879634 -0.033025833  0.03534548 -0.3722637
#>  [3,] -0.06041190 -0.01976111  0.08344769  0.088251002  0.06902883 -0.2763480
#>  [4,] -0.05836449  0.02289903 -0.01430982 -0.046441289  0.05223756 -0.3592247
#>  [5,] -0.03659839 -0.03652714 -0.03313981  0.098323705 -0.02507789 -0.2732791
#>  [6,] -0.08130015 -0.10817215 -0.11297119 -0.001842629  0.01200509 -0.1469846
#>  [7,] -0.11843234 -0.01623608  0.02735057 -0.136856192 -0.04382064 -0.2497939
#>  [8,]  0.01050287 -0.07640203  0.02089918  0.068678024  0.03334874 -0.3344031
#>  [9,] -0.07877443 -0.08165285 -0.08724109  0.032478228  0.07670425 -0.3482279
#> [10,] -0.08890671 -0.05776327  0.01126565  0.066071699  0.13185077 -0.4446857
#>               V20         V21          V22          V23         V24         V25
#>  [1,] -0.08036817 -0.04897908 -0.097820096 -0.004616038 -0.02501219 0.008009921
#>  [2,] -0.09426354 -0.02440396  0.019372925 -0.010565664  0.09317791 0.030165351
#>  [3,]  0.19661466  0.16927925  0.118833389  0.295359484  0.14172078 0.266321668
#>  [4,]  0.10298376 -0.02078917  0.024537430  0.069885629  0.12990249 0.062753477
#>  [5,]  0.02743796  0.01439067  0.119547534  0.062970194  0.21928165 0.149178064
#>  [6,] -0.01090075  0.01037883  0.172259559  0.177556142  0.06749034 0.035695345
#>  [7,]  0.03286059  0.15600866  0.069365408  0.189356489  0.13340420 0.048085362
#>  [8,]  0.08248755  0.02916602 -0.004514772  0.089573650  0.07126978 0.121601183
#>  [9,]  0.06521792  0.16019863  0.136988288  0.046147536  0.05257545 0.088721296
#> [10,]  0.17539001  0.23134814  0.261587703  0.194948693  0.31730528 0.241068804
#>                V26         V27         V28          V29         V3          V30
#>  [1,]  0.097666118 -0.04135308  0.03860431 -0.009031105 -0.2870570  0.019252787
#>  [2,]  0.037191398  0.22775019  0.06692359 -0.001016122 -0.2849909  0.071303240
#>  [3,]  0.308831074  0.21442198  0.15494462  0.188200701 -0.3553992  0.149425916
#>  [4,]  0.081906263  0.22207131  0.21629879  0.169053220 -0.2839617  0.000368969
#>  [5,]  0.091998045  0.15673716  0.10625642  0.044324403 -0.2528949  0.043080220
#>  [6,]  0.050205058  0.11831524  0.13761030  0.049446622 -0.2767547  0.053826840
#>  [7,]  0.001513002  0.08287197 -0.05712108 -0.093385948 -0.2513925 -0.108977843
#>  [8,] -0.015942016 -0.05634039 -0.08258049  0.009020242 -0.3364345  0.048480175
#>  [9,]  0.049085639  0.04375484  0.01518777  0.003934351 -0.3194725 -0.019388389
#> [10,]  0.254007835  0.30585421  0.30971387  0.243512108 -0.3267021  0.202332552
#>                V31         V32         V33         V34          V35         V36
#>  [1,] -0.097996291  0.04461038 -0.10539794  0.06906348  0.063432358 -0.03710474
#>  [2,] -0.054589919 -0.05641988 -0.03191357  0.10415534  0.037164127  0.06666467
#>  [3,]  0.010449340 -0.04064104  0.06348856  0.02026628 -0.002184712  0.06712676
#>  [4,]  0.020879057  0.11812089  0.08445424 -0.06736228 -0.098700041 -0.02124143
#>  [5,] -0.007879787 -0.02156723  0.04383888  0.04021751  0.065843326 -0.15141677
#>  [6,] -0.042946168  0.08704374  0.01305295 -0.04900883  0.084793078  0.05287311
#>  [7,]  0.040168694 -0.02070645 -0.09467700 -0.08264167 -0.112626270  0.06302131
#>  [8,] -0.023061023 -0.02929812 -0.10003905 -0.11724363 -0.076830711 -0.08082547
#>  [9,]  0.012471074 -0.02313124 -0.11439775 -0.07441281 -0.142426536 -0.06415544
#> [10,]  0.196888638  0.05569974  0.10742485  0.10423639 -0.011538445 -0.03071793
#>               V37         V38         V39         V4         V40         V41
#>  [1,] -0.08173502 -0.09190564 -0.06004608 -0.3323138  0.04534734 -0.14427224
#>  [2,]  0.06790604 -0.08323068 -0.03662931 -0.1701969  0.02987751 -0.05444241
#>  [3,]  0.11466965  0.02790657 -0.09417077 -0.2599059 -0.02537271 -0.08570493
#>  [4,] -0.05509156  0.01184496 -0.01366248 -0.3179368  0.01843116 -0.03641248
#>  [5,] -0.05024852 -0.11199233 -0.12138663 -0.1777145 -0.03979094 -0.07704441
#>  [6,] -0.09653221  0.01046914 -0.04737812 -0.2514886 -0.07975464 -0.08470972
#>  [7,] -0.06186694 -0.14730933 -0.09201553 -0.2888574 -0.06499261 -0.08431110
#>  [8,] -0.01372524 -0.08462950 -0.18458266 -0.1856657 -0.09015002 -0.15362007
#>  [9,]  0.01399599 -0.03709067 -0.09752267 -0.2722293 -0.11883306 -0.08183113
#> [10,] -0.04938331 -0.01652714  0.04725037 -0.3565349 -0.11047348  0.06432485
#>                V42         V43         V44         V45         V46         V47
#>  [1,] -0.165045927 -0.18044473 -0.06465351 -0.17922485 -0.25410843 -0.09316422
#>  [2,] -0.022880578 -0.06229796 -0.18960065 -0.17381766 -0.24540408 -0.21777985
#>  [3,] -0.045562539 -0.02520368 -0.09622522 -0.16701410 -0.15816839 -0.31217703
#>  [4,] -0.087058536 -0.11054942 -0.06000518 -0.13408559 -0.05210978 -0.10190742
#>  [5,]  0.001732069 -0.05676045 -0.19323734 -0.07718744 -0.16915835 -0.07585031
#>  [6,]  0.049999454 -0.04538292 -0.02970997 -0.11585144 -0.10060618 -0.24047961
#>  [7,] -0.136252967 -0.14536356 -0.09901393 -0.10028032 -0.11506176 -0.13038563
#>  [8,] -0.249277495 -0.18634959 -0.22310813 -0.17127607 -0.28823874 -0.22644163
#>  [9,] -0.125070022 -0.15750830 -0.21983023 -0.13060074 -0.12109044 -0.25365771
#> [10,]  0.036292758 -0.01951592 -0.11595107 -0.19229785 -0.24845802 -0.24901323
#>              V48        V49         V5        V50        V51        V52
#>  [1,] -0.1432670 -0.3479026 -0.3318114 -0.3206999 -0.3583545 -0.2343846
#>  [2,] -0.3481106 -0.3226681 -0.2575221 -0.1935213 -0.2253936 -0.3596765
#>  [3,] -0.2623143 -0.2538746 -0.2446759 -0.4462179 -0.3589005 -0.3358927
#>  [4,] -0.1675147 -0.2651882 -0.2087186 -0.3755999 -0.3722118 -0.3487798
#>  [5,] -0.2557465 -0.3621190 -0.2949675 -0.3274580 -0.2762497 -0.3391948
#>  [6,] -0.2873758 -0.2868099 -0.1620259 -0.1592297 -0.2986044 -0.3193702
#>  [7,] -0.1388615 -0.1706319 -0.1607998 -0.3069860 -0.1764247 -0.3382622
#>  [8,] -0.3559361 -0.2448674 -0.2057132 -0.3980531 -0.2328089 -0.3035129
#>  [9,] -0.2231404 -0.2567476 -0.2404624 -0.2958299 -0.3077401 -0.3786829
#> [10,] -0.1952520 -0.3317468 -0.2619381 -0.4581626 -0.4806839 -0.4178730
#>              V53        V54        V55        V56        V57        V58
#>  [1,] -0.2268854 -0.2513656 -0.3275179 -0.4020951 -0.2574958 -0.2009879
#>  [2,] -0.2367204 -0.2551815 -0.2113019 -0.3205466 -0.2861869 -0.3401492
#>  [3,] -0.4344227 -0.4098331 -0.4406741 -0.3200226 -0.4353782 -0.4191734
#>  [4,] -0.2881792 -0.3503588 -0.3601468 -0.4177999 -0.3144486 -0.4041604
#>  [5,] -0.3373726 -0.3813962 -0.3527424 -0.3916248 -0.3092069 -0.3329364
#>  [6,] -0.2766044 -0.2006657 -0.2273774 -0.3039959 -0.3661290 -0.3487242
#>  [7,] -0.2993191 -0.2712084 -0.3616620 -0.2706015 -0.3685281 -0.2487456
#>  [8,] -0.3268909 -0.2388222 -0.3580758 -0.3553050 -0.2985101 -0.2704647
#>  [9,] -0.4209708 -0.4053389 -0.3286252 -0.3789176 -0.3182954 -0.3721731
#> [10,] -0.3354333 -0.4767990 -0.4856152 -0.3623081 -0.4853108 -0.4136799
#>              V59         V6        V60          V7          V8          V9
#>  [1,] -0.3363950 -0.1411262 -0.2340067 -0.13516075 -0.14546042 -0.11562482
#>  [2,] -0.2440523 -0.2618276 -0.2888891 -0.17865762 -0.15035883 -0.16152502
#>  [3,] -0.4194227 -0.2362784 -0.4050099 -0.21675216 -0.15635513 -0.13310641
#>  [4,] -0.3948061 -0.1971205 -0.2559804 -0.09932173 -0.13174492 -0.16402614
#>  [5,] -0.2730879 -0.2687464 -0.2951039 -0.17930879 -0.24614079 -0.09935377
#>  [6,] -0.2713530 -0.1890372 -0.3104594 -0.15386903 -0.08420163 -0.21234308
#>  [7,] -0.3753862 -0.1404684 -0.2968928 -0.21416187 -0.12733279 -0.08881712
#>  [8,] -0.3727610 -0.2675233 -0.3186321 -0.19936382 -0.23974977 -0.16322297
#>  [9,] -0.3543068 -0.1996202 -0.3220876 -0.21776522 -0.13711540 -0.26807595
#> [10,] -0.3347960 -0.2315084 -0.4735281 -0.27376451 -0.18268787 -0.11021531
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.18858868 -0.1431542 -0.06911076 -0.14246363 -0.12191851 -0.08073056
#> [2,]  0.08625959  0.1076164 -0.05347435  0.08325328  0.02347484  0.05706580
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.02211801 -0.11015967 -0.02368063 0.042900510
#> [2,] -0.03803665  0.09806508  0.07830778 0.001951697
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.029419e-04  0.0012235432  0.0015249150  0.0011138803  0.0006724971
#>  [2,]  7.597291e-05  0.0009178723  0.0011167382  0.0008134495  0.0005229622
#>  [3,] -4.704896e-05 -0.0004430242 -0.0005387191 -0.0003103125 -0.0001861977
#>  [4,]  1.018282e-04  0.0009573571  0.0011530242  0.0008899537  0.0006648333
#>  [5,]  4.068733e-05  0.0003770370  0.0004448754  0.0003789136  0.0003215228
#>  [6,]  3.131664e-05  0.0003012787  0.0003513474  0.0003052834  0.0002637663
#>  [7,] -6.131536e-05 -0.0006826363 -0.0008794645 -0.0006416634 -0.0003518377
#>  [8,]  6.960901e-05  0.0008099973  0.0010311172  0.0007995905  0.0005010657
#>  [9,]  2.144198e-05  0.0002080173  0.0002436928  0.0002237930  0.0002021364
#> [10,] -7.637663e-05 -0.0006215762 -0.0007417615 -0.0004554041 -0.0003196593
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0002365936  3.367998e-05 -3.396212e-04 -3.586997e-04 -6.515947e-04
#>  [2,]  0.0002545198  1.223988e-04 -1.267778e-04 -1.359810e-04 -3.466898e-04
#>  [3,] -0.0001760075 -1.475208e-04 -4.573667e-05 -2.055544e-05  1.270889e-04
#>  [4,]  0.0004629464  3.946305e-04  2.242070e-04  2.276397e-04  9.632725e-05
#>  [5,]  0.0002461682  2.382098e-04  2.059589e-04  2.332586e-04  2.183119e-04
#>  [6,]  0.0002021537  1.958305e-04  1.758980e-04  2.022882e-04  1.973248e-04
#>  [7,] -0.0000787932  5.541899e-05  3.039615e-04  3.520535e-04  5.543775e-04
#>  [8,]  0.0001748815  6.298654e-05 -1.795547e-04 -1.795440e-04 -3.791771e-04
#>  [9,]  0.0001541311  1.554196e-04  1.526343e-04  1.814158e-04  1.880117e-04
#> [10,] -0.0003509214 -3.586705e-04 -2.473287e-04 -2.139850e-04 -6.980524e-05
#>                 V19            V2           V20           V21           V22
#>  [1,] -3.015338e-04  1.486198e-04 -7.086344e-05  1.067132e-04  6.087926e-05
#>  [2,] -1.085441e-04  1.030207e-04  8.845441e-05  2.486882e-04  2.402322e-04
#>  [3,]  3.331616e-05 -6.445388e-05 -7.260504e-05 -1.231281e-04 -6.980424e-06
#>  [4,]  4.049012e-04  1.458439e-04  7.071491e-04  8.636704e-04  7.309847e-04
#>  [5,]  3.353068e-04  5.915914e-05  4.720120e-04  5.451434e-04  5.167764e-04
#>  [6,]  2.871109e-04  4.477784e-05  3.941172e-04  4.532022e-04  4.421940e-04
#>  [7,]  3.193197e-04 -9.215583e-05  1.942116e-04  9.232814e-05  1.548175e-04
#>  [8,] -1.421442e-04  1.008080e-04 -8.796243e-06  1.116744e-04  9.678439e-05
#>  [9,]  2.500123e-04  3.131986e-05  3.253641e-04  3.681533e-04  3.675853e-04
#> [10,] -2.030347e-04 -1.020835e-04 -3.783305e-04 -4.221033e-04 -1.990740e-04
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0007028580 -9.875869e-04 -9.835012e-04 -0.0019085116 -0.0021626549
#>  [2,] -0.0002464369 -3.937055e-04 -3.505503e-04 -0.0009555130 -0.0011163276
#>  [3,]  0.0003354373  4.337453e-04  3.122796e-04  0.0006568844  0.0009041443
#>  [4,]  0.0002075559  8.506976e-05  3.817329e-05 -0.0004660818 -0.0006393892
#>  [5,]  0.0003436268  3.212801e-04  2.849211e-04  0.0001486840  0.0001142478
#>  [6,]  0.0003180219  3.016767e-04  2.687883e-04  0.0001684960  0.0001557900
#>  [7,]  0.0006137199  7.953778e-04  8.053259e-04  0.0013714954  0.0015310068
#>  [8,] -0.0003972969 -6.199770e-04 -6.660711e-04 -0.0012763010 -0.0013822654
#>  [9,]  0.0002879478  2.772131e-04  2.436809e-04  0.0001880677  0.0001908472
#> [10,]  0.0002812746  4.152557e-04  2.916091e-04  0.0007177698  0.0010430750
#>                 V28           V29            V3           V30           V31
#>  [1,] -0.0022189743 -2.602182e-03  1.851479e-04 -2.799580e-03 -3.158265e-03
#>  [2,] -0.0011582198 -1.455579e-03  1.364739e-04 -1.634469e-03 -1.921501e-03
#>  [3,]  0.0009972861  1.138141e-03 -8.236422e-05  1.098591e-03  1.250672e-03
#>  [4,] -0.0006773145 -8.587439e-04  1.732153e-04 -9.839510e-04 -1.326547e-03
#>  [5,]  0.0001006348  5.122221e-05  6.948111e-05 -1.865815e-05 -1.669066e-04
#>  [6,]  0.0001524521  1.093465e-04  5.307410e-05  4.377237e-05 -7.003266e-05
#>  [7,]  0.0015539755  1.777957e-03 -1.106246e-04  1.874978e-03  2.064167e-03
#>  [8,] -0.0013984021 -1.659808e-03  1.259788e-04 -1.822411e-03 -2.063644e-03
#>  [9,]  0.0001879721  1.606231e-04  3.673013e-05  1.059827e-04  2.106776e-05
#> [10,]  0.0011378811  1.267397e-03 -1.241559e-04  1.180350e-03  1.420458e-03
#>                 V32           V33           V34           V35           V36
#>  [1,] -2.278118e-03 -2.753285e-03 -3.619646e-03 -3.966651e-03 -3.941749e-03
#>  [2,] -1.351915e-03 -1.671244e-03 -2.258099e-03 -2.491861e-03 -2.468367e-03
#>  [3,]  8.968873e-04  1.040294e-03  1.322532e-03  1.439549e-03  1.513562e-03
#>  [4,] -8.628852e-04 -1.146779e-03 -1.569704e-03 -1.760404e-03 -1.857567e-03
#>  [5,] -5.841177e-05 -1.599565e-04 -2.926514e-04 -3.495599e-04 -3.840988e-04
#>  [6,]  5.217910e-06 -7.346471e-05 -1.801136e-04 -2.172826e-04 -2.366195e-04
#>  [7,]  1.518536e-03  1.809338e-03  2.291911e-03  2.530829e-03  2.536992e-03
#>  [8,] -1.493744e-03 -1.799353e-03 -2.349258e-03 -2.564317e-03 -2.553266e-03
#>  [9,]  5.589069e-05  2.543605e-06 -6.332828e-05 -8.441262e-05 -9.325644e-05
#> [10,]  9.999748e-04  1.139751e-03  1.437148e-03  1.554862e-03  1.717788e-03
#>                 V37           V38           V39            V4           V40
#>  [1,] -2.806069e-03 -1.683414e-03 -7.609246e-04  1.862877e-04 -1.710039e-03
#>  [2,] -1.728763e-03 -1.015896e-03 -3.685864e-04  1.370015e-04 -1.032775e-03
#>  [3,]  1.139490e-03  7.300532e-04  4.590711e-04 -1.099153e-04  8.107999e-04
#>  [4,] -1.268240e-03 -5.934684e-04 -6.473011e-05  1.853087e-04 -6.857606e-04
#>  [5,] -2.136897e-04 -3.111751e-05  1.214630e-04  6.894884e-05 -6.383773e-05
#>  [6,] -1.116434e-04  1.095050e-05  1.254022e-04  5.098042e-05 -9.159110e-06
#>  [7,]  1.853206e-03  1.103072e-03  5.079168e-04 -1.192789e-04  1.104376e-03
#>  [8,] -1.802331e-03 -1.069266e-03 -4.406165e-04  1.206280e-04 -1.060450e-03
#>  [9,] -1.480257e-05  4.959025e-05  1.171922e-04  3.409311e-05  3.620398e-05
#> [10,]  1.298383e-03  7.765219e-04  4.673628e-04 -1.642834e-04  9.193498e-04
#>                 V41           V42           V43           V44           V45
#>  [1,] -1.420981e-03 -8.753280e-04 -3.703049e-04  2.641004e-04  0.0007485481
#>  [2,] -8.943343e-04 -5.427909e-04 -1.806992e-04  2.619489e-04  0.0005646299
#>  [3,]  6.259842e-04  4.052448e-04  1.874247e-04 -3.372667e-05 -0.0002410025
#>  [4,] -5.324743e-04 -2.066500e-04  5.725162e-05  4.178200e-04  0.0007426526
#>  [5,] -4.010837e-05  4.988932e-05  1.079697e-04  2.225244e-04  0.0003216726
#>  [6,] -7.447822e-06  5.276025e-05  9.414377e-05  1.813184e-04  0.0002463428
#>  [7,]  8.808027e-04  5.204189e-04  2.149111e-04 -1.394589e-04 -0.0004657248
#>  [8,] -8.832093e-04 -5.245437e-04 -2.179889e-04  1.813898e-04  0.0004855043
#>  [9,]  2.919261e-05  6.393848e-05  8.185500e-05  1.355469e-04  0.0001731643
#> [10,]  7.239491e-04  4.642775e-04  1.826538e-04 -8.319396e-05 -0.0003355269
#>                 V46           V47           V48           V49            V5
#>  [1,]  0.0004102777  1.965839e-04  2.432161e-04  8.182725e-05  1.003831e-04
#>  [2,]  0.0002937335  1.407075e-04  1.847375e-04  6.333856e-05  7.967236e-05
#>  [3,] -0.0001338430 -1.181719e-05 -6.218654e-05 -1.564048e-05 -7.198169e-05
#>  [4,]  0.0004771236  2.646207e-04  2.588797e-04  1.096208e-04  1.363101e-04
#>  [5,]  0.0002227646  1.468939e-04  1.254588e-04  5.716068e-05  5.826985e-05
#>  [6,]  0.0001690220  1.165544e-04  9.929489e-05  4.506508e-05  4.495671e-05
#>  [7,] -0.0002628613 -1.158849e-04 -1.413845e-04 -4.955816e-05 -6.660264e-05
#>  [8,]  0.0002559446  1.401434e-04  1.641695e-04  5.521737e-05  7.425834e-05
#>  [9,]  0.0001221337  9.053248e-05  7.346249e-05  3.413351e-05  3.195172e-05
#> [10,] -0.0001923637 -3.058078e-05 -8.872105e-05 -2.555223e-05 -1.183058e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -5.668804e-05 -8.034888e-06  9.737618e-06 -2.110625e-05 -5.141761e-06
#>  [2,] -3.661744e-05 -4.831788e-06  8.077561e-06 -1.164408e-05 -4.567997e-07
#>  [3,]  1.862931e-05  1.334278e-06 -4.038793e-06  3.767441e-06 -1.139204e-07
#>  [4,] -8.552318e-06  1.492034e-05  2.621192e-05  2.366508e-06  1.191841e-05
#>  [5,]  5.476111e-06  1.025368e-05  1.413499e-05  4.073244e-06  6.679743e-06
#>  [6,]  4.910990e-06  7.726544e-06  1.102807e-05  3.595603e-06  5.344485e-06
#>  [7,]  3.526225e-05  2.096411e-06 -5.320968e-06  1.293672e-05  1.223811e-06
#>  [8,] -3.817138e-05 -4.686981e-06  6.917291e-06 -1.278094e-05 -1.605188e-06
#>  [9,]  4.930371e-06  6.128625e-06  8.396008e-06  3.138443e-06  4.037418e-06
#> [10,]  1.454376e-05 -1.924040e-06 -1.079702e-05 -1.849573e-06 -4.766689e-06
#>                 V55           V56           V57           V58           V59
#>  [1,] -5.118647e-06  6.438555e-06 -2.756398e-05  2.566355e-05  6.229055e-06
#>  [2,] -2.170022e-07  7.691001e-06 -1.528527e-05  1.908771e-05  6.065178e-06
#>  [3,]  9.179238e-07 -5.207284e-06  5.773118e-06 -8.976903e-06 -2.672071e-06
#>  [4,]  1.159929e-05  1.529922e-05 -9.442225e-06  2.455314e-05  1.414547e-05
#>  [5,]  7.541226e-06  7.831750e-06 -1.070125e-06  9.764348e-06  6.586154e-06
#>  [6,]  6.121015e-06  6.453253e-06 -3.197739e-07  7.737740e-06  5.196986e-06
#>  [7,]  3.609691e-06 -2.401431e-06  1.812232e-05 -1.624765e-05 -5.394175e-06
#>  [8,] -2.020958e-06  5.189940e-06 -1.831013e-05  1.888058e-05  5.945798e-06
#>  [9,]  4.778740e-06  4.765025e-06  1.676046e-07  5.452964e-06  3.791100e-06
#> [10,] -3.887232e-06 -1.107535e-05  3.995040e-06 -1.395639e-05 -6.364632e-06
#>                  V6           V60            V7            V8            V9
#>  [1,] -3.199028e-05  2.905288e-06 -1.256109e-04  1.674746e-04  0.0009184384
#>  [2,]  5.666221e-06  3.871222e-06 -4.722116e-05  1.668417e-04  0.0007031280
#>  [3,]  2.039638e-05 -4.957303e-06  4.409491e-05 -5.086926e-05 -0.0003476117
#>  [4,]  8.049022e-05  1.006422e-05  1.851451e-05  2.165681e-04  0.0007560349
#>  [5,]  6.548400e-05  4.793802e-06  4.334554e-05  1.097348e-04  0.0002990001
#>  [6,]  5.752966e-05  3.791659e-06  4.196624e-05  9.334117e-05  0.0002372362
#>  [7,]  3.628387e-05 -1.857650e-06  1.004226e-04 -7.673610e-05 -0.0005170635
#>  [8,] -3.597193e-06  1.818455e-06 -7.418170e-05  1.114012e-04  0.0005875061
#>  [9,]  4.691658e-05  2.606610e-06  3.509979e-05  6.852545e-05  0.0001629101
#> [10,] -7.013538e-06 -9.640755e-06  1.930926e-05 -9.337700e-05 -0.0004831167
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.04803371  0.06319326  0.07181141  0.06147485  0.05928561  0.05889419
#> [2,] -0.04803371 -0.06319326 -0.07181141 -0.06147485 -0.05928561 -0.05889419
#>             [,7]       [,8]       [,9]      [,10]
#> [1,]  0.04368003  0.0422127  0.0456537  0.0744829
#> [2,] -0.04368003 -0.0422127 -0.0456537 -0.0744829
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.2950747 0.3542896 0.8009996 0.4469079 0.4945614 0.3536025 0.3113696
#>  [8] 0.3280676 0.3875999 0.7177533
#> 
#> $B[[2]]
#> [1]  0.07642965 -0.06205901
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -0.0036371911 -0.0020798713  0.0013686385 -0.0010468036  0.0001078322
#>  [6]  0.0001685595  0.0024649763 -0.0023081903  0.0002304031  0.0013601690
#> 
#> $vB[[2]]
#> [1]  0.07927793 -0.07927793
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#>  [2,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#>  [3,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#>  [4,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#>  [5,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#>  [6,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#>  [7,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#>  [8,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#>  [9,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [10,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [11,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [12,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [13,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [14,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [15,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [16,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [17,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [18,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [19,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [20,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [21,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [22,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [23,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [24,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [25,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [26,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [27,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [28,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [29,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [30,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [31,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [32,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#> [33,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [34,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [35,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [36,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [37,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [38,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [39,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [40,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#>  [2,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#>  [3,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#>  [4,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#>  [5,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#>  [6,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#>  [7,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#>  [8,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#>  [9,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [10,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [11,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [12,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [13,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [14,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [15,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [16,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [17,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [18,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [19,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [20,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [21,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [22,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [23,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [24,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [25,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [26,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [27,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [28,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [29,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [30,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [31,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [32,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#> [33,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [34,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [35,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [36,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [37,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [38,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [39,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [40,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#>  [2,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#>  [3,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#>  [4,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#>  [5,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#>  [6,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#>  [7,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#>  [8,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#>  [9,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [10,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [11,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [12,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [13,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [14,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [15,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [16,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [17,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [18,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [19,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [20,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [21,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [22,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [23,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [24,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [25,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [26,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [27,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [28,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [29,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [30,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [31,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [32,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#> [33,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [34,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [35,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [36,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [37,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [38,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [39,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [40,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#>  [2,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#>  [3,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#>  [4,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#>  [5,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#>  [6,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#>  [7,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#>  [8,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#>  [9,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [10,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [11,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [12,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [13,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [14,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [15,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [16,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [17,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [18,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [19,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [20,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [21,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [22,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [23,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [24,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [25,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [26,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [27,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [28,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [29,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [30,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [31,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [32,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#> [33,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [34,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [35,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [36,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [37,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [38,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [39,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [40,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#>  [2,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#>  [3,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#>  [4,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#>  [5,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#>  [6,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#>  [7,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#>  [8,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#>  [9,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [10,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [11,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [12,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [13,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [14,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [15,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [16,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [17,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [18,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [19,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [20,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [21,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [22,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [23,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [24,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [25,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [26,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [27,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [28,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [29,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [30,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [31,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [32,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#> [33,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [34,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [35,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [36,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [37,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [38,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [39,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [40,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#>  [2,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#>  [3,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#>  [4,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#>  [5,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#>  [6,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#>  [7,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#>  [8,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#>  [9,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [10,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [11,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [12,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [13,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [14,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [15,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [16,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [17,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [18,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [19,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [20,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [21,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [22,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [23,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [24,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [25,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [26,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [27,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [28,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [29,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [30,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [31,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [32,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#> [33,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [34,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [35,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [36,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [37,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [38,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [39,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [40,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4497348 0.5649231 0.8960724 0.7518291 0.7158986 0.6735038 0.5202825
#>  [2,] 0.3953660 0.5371921 0.7723541 0.6016141 0.5305892 0.5641528 0.3553860
#>  [3,] 0.4180037 0.5322104 0.8586563 0.7054116 0.6675373 0.6145695 0.4303685
#>  [4,] 0.5029995 0.5945099 0.8585218 0.6985398 0.6965919 0.6278550 0.5180847
#>  [5,] 0.3743245 0.4781274 0.8902908 0.6882640 0.6619890 0.6153131 0.4843675
#>  [6,] 0.4022825 0.4972066 0.8794040 0.6285968 0.6402579 0.5789734 0.5141490
#>  [7,] 0.3492446 0.4946499 0.8898024 0.6658773 0.6429265 0.6113727 0.4359233
#>  [8,] 0.4147921 0.5589210 0.7998995 0.6333860 0.5688090 0.5887711 0.3682200
#>  [9,] 0.4484114 0.5485996 0.8665892 0.7138571 0.6821264 0.6267095 0.4531193
#> [10,] 0.3764891 0.4780903 0.8706065 0.6680806 0.6423178 0.6390046 0.4852089
#> [11,] 0.4602293 0.5549427 0.9160613 0.7263438 0.7439006 0.6603565 0.5893242
#> [12,] 0.4788686 0.5818004 0.9329281 0.7625751 0.7615788 0.6775673 0.5789029
#> [13,] 0.3515562 0.5132037 0.8270357 0.6191121 0.5870406 0.5815612 0.3644290
#> [14,] 0.4541269 0.5723242 0.8900322 0.7397539 0.7124820 0.6552757 0.4949299
#> [15,] 0.3132496 0.4264880 0.8713758 0.6769414 0.6014176 0.5814687 0.4027254
#> [16,] 0.4165012 0.5551639 0.7810816 0.6113160 0.5474282 0.5751185 0.3819614
#> [17,] 0.3969999 0.3733088 0.7263122 0.5066774 0.5205402 0.4430398 0.3845525
#> [18,] 0.3880578 0.4308895 0.8814528 0.6551453 0.6476808 0.5462766 0.4477827
#> [19,] 0.4023614 0.5274021 0.8771915 0.6983633 0.6773546 0.6472759 0.5225730
#> [20,] 0.4613670 0.5961291 0.8745450 0.6981781 0.6775846 0.6361189 0.4996173
#> [21,] 0.4334185 0.5529972 0.8988515 0.7298480 0.7088094 0.6573276 0.5296034
#> [22,] 0.3549551 0.3558648 0.7669700 0.5321404 0.5371902 0.4757396 0.4081093
#> [23,] 0.4404502 0.6052047 0.9095483 0.7315324 0.6975237 0.6776789 0.4979710
#> [24,] 0.4161811 0.5476883 0.7620919 0.5895113 0.5409639 0.5672387 0.3721903
#> [25,] 0.4258797 0.5417604 0.8580240 0.6969158 0.6553491 0.6319323 0.4389873
#> [26,] 0.4424503 0.5492660 0.7971671 0.6559730 0.5973060 0.5792629 0.4215337
#> [27,] 0.4074113 0.4887339 0.8929446 0.7081123 0.6932609 0.6274319 0.5109503
#> [28,] 0.3343831 0.4513130 0.7601639 0.5676025 0.5240375 0.5159899 0.3583368
#> [29,] 0.3854441 0.4657389 0.8976310 0.6822774 0.6699796 0.5999837 0.5102444
#> [30,] 0.3661625 0.4619816 0.8799456 0.6715930 0.6484927 0.5904420 0.4768000
#> [31,] 0.3863174 0.4945003 0.8927311 0.6565263 0.6617362 0.6118325 0.4982431
#> [32,] 0.3721367 0.5256230 0.8999195 0.6671232 0.6746551 0.6436461 0.4770978
#> [33,] 0.3426875 0.4579495 0.8309137 0.5670498 0.5913016 0.5538224 0.4206143
#> [34,] 0.3471027 0.4649371 0.8388287 0.6050323 0.5963021 0.5580934 0.4459413
#> [35,] 0.5028164 0.6163691 0.8799434 0.7316844 0.7182580 0.6718036 0.5065669
#> [36,] 0.3559720 0.5009158 0.8483484 0.6695032 0.6057899 0.5913049 0.3589500
#> [37,] 0.3698929 0.4119282 0.8039724 0.5469401 0.5643984 0.5038819 0.4250532
#> [38,] 0.2752407 0.3590897 0.7424962 0.4992461 0.4330424 0.4442905 0.2662383
#> [39,] 0.3995170 0.4516252 0.8924282 0.6681534 0.6783374 0.5706587 0.4408247
#> [40,] 0.4359956 0.5610824 0.9053763 0.7390474 0.7178698 0.6610558 0.5023508
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4466444 0.5599327 0.9363751
#>  [2,] 0.2719183 0.3673109 0.8205298
#>  [3,] 0.3701346 0.4868795 0.9085574
#>  [4,] 0.4703092 0.5453856 0.8866681
#>  [5,] 0.3636257 0.5065366 0.9181098
#>  [6,] 0.4542569 0.5627050 0.8901734
#>  [7,] 0.3315238 0.4787671 0.9232670
#>  [8,] 0.2846068 0.3789921 0.8556372
#>  [9,] 0.4100483 0.5022426 0.9143296
#> [10,] 0.4086215 0.5157459 0.9152258
#> [11,] 0.5451461 0.6275706 0.9448282
#> [12,] 0.5501704 0.6314749 0.9533256
#> [13,] 0.2672138 0.3831556 0.8693636
#> [14,] 0.4300409 0.5412057 0.9346983
#> [15,] 0.2743452 0.4357095 0.9196121
#> [16,] 0.2991944 0.3930199 0.8241046
#> [17,] 0.3864709 0.4365448 0.7280318
#> [18,] 0.4030580 0.4940743 0.9118578
#> [19,] 0.4257862 0.5726575 0.9167391
#> [20,] 0.4217848 0.5271876 0.9056905
#> [21,] 0.4561886 0.5777858 0.9315714
#> [22,] 0.3740017 0.4457363 0.7774883
#> [23,] 0.4178938 0.5354492 0.9384900
#> [24,] 0.3000906 0.3945204 0.8029274
#> [25,] 0.3968438 0.4786673 0.9040902
#> [26,] 0.3975836 0.4788203 0.8452358
#> [27,] 0.4589409 0.5589516 0.9286769
#> [28,] 0.2565847 0.3577824 0.8030766
#> [29,] 0.4396649 0.5566837 0.9251434
#> [30,] 0.3482355 0.4898152 0.9096447
#> [31,] 0.4385785 0.5401656 0.9168228
#> [32,] 0.3697087 0.5049068 0.9264462
#> [33,] 0.3093461 0.4256973 0.8640390
#> [34,] 0.3626368 0.4627166 0.8584530
#> [35,] 0.4546003 0.5296422 0.9205813
#> [36,] 0.2601312 0.3712435 0.8984180
#> [37,] 0.3979698 0.4748177 0.8212862
#> [38,] 0.2036583 0.3085791 0.7796827
#> [39,] 0.4247480 0.5106623 0.9230380
#> [40,] 0.4375357 0.5573117 0.9413913
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5796673 0.4203327
#>  [2,] 0.5807848 0.4192152
#>  [3,] 0.5815867 0.4184133
#>  [4,] 0.5739206 0.4260794
#>  [5,] 0.5884977 0.4115023
#>  [6,] 0.5847699 0.4152301
#>  [7,] 0.5898299 0.4101701
#>  [8,] 0.5805082 0.4194918
#>  [9,] 0.5792437 0.4207563
#> [10,] 0.5868249 0.4131751
#> [11,] 0.5796365 0.4203635
#> [12,] 0.5773975 0.4226025
#> [13,] 0.5868046 0.4131954
#> [14,] 0.5793109 0.4206891
#> [15,] 0.5938268 0.4061732
#> [16,] 0.5791300 0.4208700
#> [17,] 0.5800964 0.4199036
#> [18,] 0.5879511 0.4120489
#> [19,] 0.5837374 0.4162626
#> [20,] 0.5782141 0.4217859
#> [21,] 0.5811789 0.4188211
#> [22,] 0.5860385 0.4139615
#> [23,] 0.5803464 0.4196536
#> [24,] 0.5781476 0.4218524
#> [25,] 0.5803692 0.4196308
#> [26,] 0.5759939 0.4240061
#> [27,] 0.5842233 0.4157767
#> [28,] 0.5865315 0.4134685
#> [29,] 0.5873491 0.4126509
#> [30,] 0.5895655 0.4104345
#> [31,] 0.5863213 0.4136787
#> [32,] 0.5877900 0.4122100
#> [33,] 0.5901627 0.4098373
#> [34,] 0.5876918 0.4123082
#> [35,] 0.5745447 0.4254553
#> [36,] 0.5876642 0.4123358
#> [37,] 0.5858124 0.4141876
#> [38,] 0.5922928 0.4077072
#> [39,] 0.5861747 0.4138253
#> [40,] 0.5811020 0.4188980
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]         [,2]      [,3]         [,4]        [,5]        [,6]
#>  [1,] -0.20174238  0.261166625 2.1543269  1.108391643  0.92420743  0.72407534
#>  [2,] -0.42481134  0.149043795 1.2216519  0.412194988  0.12250986  0.25803337
#>  [3,] -0.33097362  0.129020197 1.8041740  0.873202537  0.69706765  0.46656099
#>  [4,]  0.01199826  0.382641042 1.8030667  0.840354146  0.83112091  0.52302586
#>  [5,] -0.51370895 -0.087546118 2.0937147  0.792015662  0.67217065  0.46970163
#>  [6,] -0.39596382 -0.011173528 1.9867979  0.526201559  0.57648376  0.31856065
#>  [7,] -0.62236133 -0.021401236 2.0887247  0.689597213  0.58808892  0.45308614
#>  [8,] -0.34418978  0.236784227 1.3856662  0.546770638  0.27699336  0.35888732
#>  [9,] -0.20709131  0.195014075 1.8711323  0.914191463  0.76356130  0.51812619
#> [10,] -0.50447714 -0.087695102 1.9063322  0.699516596  0.58543851  0.57104659
#> [11,] -0.15941975  0.220661900 2.3899966  0.976151019  1.06634181  0.66488318
#> [12,] -0.08457589  0.330168451 2.6325634  1.166849419  1.16135464  0.74261453
#> [13,] -0.61220585  0.052827026 1.5647625  0.485781442  0.35174482  0.32918556
#> [14,] -0.18400990  0.291340374 2.0910698  1.044689687  0.90746904  0.64231094
#> [15,] -0.78497046 -0.296194808 1.9131787  0.739750799  0.41137547  0.32880546
#> [16,] -0.33715288  0.221557334 1.2719806  0.452847242  0.19028486  0.30276566
#> [17,] -0.41798136 -0.518048165 0.9759921  0.026711136  0.08220689 -0.22883411
#> [18,] -0.45548338 -0.278222877 2.0062607  0.641733745  0.60886019  0.18563765
#> [19,] -0.39563555  0.109718324 1.9660989  0.839515999  0.74164089  0.60708667
#> [20,] -0.15484065  0.389361802 1.9417566  0.838637001  0.74269355  0.55855831
#> [21,] -0.26791708  0.212787884 2.1845283  0.993851338  0.88960861  0.65140766
#> [22,] -0.59732881 -0.593357653 1.1912807  0.128739231  0.14903610 -0.09711780
#> [23,] -0.23933506  0.427199289 2.3081324  1.002411523  0.83553354  0.74312555
#> [24,] -0.33847016  0.191334672 1.1641828  0.361945548  0.16422358  0.27059387
#> [25,] -0.29868210  0.167431635 1.7989741  0.832653923  0.64263616  0.54051536
#> [26,] -0.23122333  0.197705468 1.3686820  0.645399419  0.39425275  0.31974823
#> [27,] -0.37467743 -0.045071850 2.1211778  0.886233627  0.81540901  0.52121524
#> [28,] -0.68842708 -0.195367104 1.1535783  0.272076115  0.09622424  0.06398135
#> [29,] -0.46650356 -0.137259530 2.1711755  0.764257687  0.70809265  0.40539715
#> [30,] -0.54871529 -0.152367518 1.9919151  0.715398785  0.61242040  0.36579304
#> [31,] -0.46281854 -0.021999833 2.1189466  0.647852307  0.67104084  0.45502150
#> [32,] -0.52306131  0.102581851 2.1963304  0.695202288  0.72931591  0.59122481
#> [33,] -0.65134078 -0.168600295 1.5921165  0.269824342  0.36934883  0.21612712
#> [34,] -0.63179901 -0.140482195 1.6495385  0.426477918  0.39008065  0.23342783
#> [35,]  0.01126561  0.474164945 1.9918943  1.003185273  0.93583713  0.71635393
#> [36,] -0.59289025  0.003663353 1.7217057  0.705939181  0.42964931  0.36936246
#> [37,] -0.53267627 -0.356000074 1.4113097  0.188314906  0.25903219  0.01552807
#> [38,] -0.96819352 -0.579317482 1.0589835 -0.003015526 -0.26944900 -0.22376724
#> [39,] -0.40747796 -0.194106482 2.1157878  0.699844930  0.74614146  0.28453900
#> [40,] -0.25742981  0.245556246 2.2584426  1.041023264  0.93391970  0.66800286
#>               [,7]       [,8]         [,9]     [,10]
#>  [1,]  0.081174401 -0.2142383  0.240889041 2.6890114
#>  [2,] -0.595447166 -0.9849115 -0.543770318 1.5199413
#>  [3,] -0.280347772 -0.5316396 -0.052494237 2.2961469
#>  [4,]  0.072370192 -0.1189030  0.182043488 2.0571499
#>  [5,] -0.062550491 -0.5596620  0.026147781 2.4169381
#>  [6,]  0.056611309 -0.1834853  0.252147398 2.0925135
#>  [7,] -0.257724002 -0.7013013 -0.084982788 2.4875865
#>  [8,] -0.539860686 -0.9217236 -0.493828561 1.7795169
#>  [9,] -0.188075331 -0.3637659  0.008970346 2.3676841
#> [10,] -0.059181471 -0.3696672  0.063004294 2.3791796
#> [11,]  0.361172243  0.1810776  0.521808873 2.8405505
#> [12,]  0.318271133  0.2013594  0.538549254 3.0167599
#> [13,] -0.556191864 -1.0088049 -0.476175514 1.8953433
#> [14,] -0.020281112 -0.2816844  0.165197370 2.6612051
#> [15,] -0.394122030 -0.9726873 -0.258593465 2.4370879
#> [16,] -0.481231415 -0.8511371 -0.434635610 1.5444083
#> [17,] -0.470269297 -0.4621713 -0.255196729 0.9846594
#> [18,] -0.209633450 -0.3927397 -0.023703823 2.3365328
#> [19,]  0.090353581 -0.2990643  0.292702120 2.3988441
#> [20,] -0.001530625 -0.3154511  0.108857712 2.2621154
#> [21,]  0.118552368 -0.1756961  0.313690414 2.6110812
#> [22,] -0.371787022 -0.5150873 -0.217913083 1.2510890
#> [23,] -0.008116159 -0.3314255  0.142035291 2.7250728
#> [24,] -0.522831666 -0.8468663 -0.428349971 1.4046924
#> [25,] -0.245273078 -0.4186333 -0.085382439 2.2435205
#> [26,] -0.316480709 -0.4155439 -0.084769465 1.6977126
#> [27,]  0.043808321 -0.1646069  0.236908358 2.5665402
#> [28,] -0.582590235 -1.0637958 -0.585002083 1.4056351
#> [29,]  0.040983494 -0.2425221  0.227713746 2.5143751
#> [30,] -0.092866806 -0.6268042 -0.040744793 2.3093039
#> [31,] -0.007027553 -0.2469332  0.161009220 2.3999408
#> [32,] -0.091672768 -0.5334668  0.019627795 2.5333388
#> [33,] -0.320252281 -0.8031781 -0.299427919 1.8492501
#> [34,] -0.217083254 -0.5639379 -0.149411008 1.8024999
#> [35,]  0.026269159 -0.1821002  0.118707988 2.4502718
#> [36,] -0.579924302 -1.0452866 -0.526886017 2.1797689
#> [37,] -0.302063216 -0.4139315 -0.100814594 1.5250861
#> [38,] -1.013792584 -1.3635848 -0.806770459 1.2638182
#> [39,] -0.237815879 -0.3033120  0.042655632 2.4843591
#> [40,]  0.009403453 -0.2511694  0.230258920 2.7764752
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] 0.03239645 -0.2890115
#>  [2,] 0.05856735 -0.2674285
#>  [3,] 0.04583565 -0.2834548
#>  [4,] 0.02115473 -0.2767106
#>  [5,] 0.06126480 -0.2964936
#>  [6,] 0.05520610 -0.2871795
#>  [7,] 0.06769233 -0.2955700
#>  [8,] 0.05246156 -0.2723985
#>  [9,] 0.03720346 -0.2824658
#> [10,] 0.06008435 -0.2907705
#> [11,] 0.02964493 -0.2916365
#> [12,] 0.02102283 -0.2910762
#> [13,] 0.06805293 -0.2827183
#> [14,] 0.03273417 -0.2872108
#> [15,] 0.08230305 -0.2975048
#> [16,] 0.05180757 -0.2673954
#> [17,] 0.06578944 -0.2573797
#> [18,] 0.06369788 -0.2918040
#> [19,] 0.04876252 -0.2893722
#> [20,] 0.03269146 -0.2827549
#> [21,] 0.03760925 -0.2900056
#> [22,] 0.07674710 -0.2708656
#> [23,] 0.03440310 -0.2897924
#> [24,] 0.05201745 -0.2631561
#> [25,] 0.04270182 -0.2815873
#> [26,] 0.04037788 -0.2659714
#> [27,] 0.04843148 -0.2917035
#> [28,] 0.07703688 -0.2726082
#> [29,] 0.05835025 -0.2946673
#> [30,] 0.06593511 -0.2962343
#> [31,] 0.05673299 -0.2920455
#> [32,] 0.05847334 -0.2963635
#> [33,] 0.07694189 -0.2876961
#> [34,] 0.06938420 -0.2850475
#> [35,] 0.01894336 -0.2814748
#> [36,] 0.06651603 -0.2878015
#> [37,] 0.07126424 -0.2754165
#> [38,] 0.10326038 -0.2701913
#> [39,] 0.05700585 -0.2911682
#> [40,] 0.03660680 -0.2906919
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4203327 -0.4203327
#>  [2,] -0.5807848  0.5807848
#>  [3,]  0.4184133 -0.4184133
#>  [4,] -0.5739206  0.5739206
#>  [5,]  0.4115023 -0.4115023
#>  [6,] -0.5847699  0.5847699
#>  [7,]  0.4101701 -0.4101701
#>  [8,] -0.5805082  0.5805082
#>  [9,] -0.5792437  0.5792437
#> [10,] -0.5868249  0.5868249
#> [11,] -0.5796365  0.5796365
#> [12,] -0.5773975  0.5773975
#> [13,] -0.5868046  0.5868046
#> [14,]  0.4206891 -0.4206891
#> [15,]  0.4061732 -0.4061732
#> [16,] -0.5791300  0.5791300
#> [17,]  0.4199036 -0.4199036
#> [18,] -0.5879511  0.5879511
#> [19,] -0.5837374  0.5837374
#> [20,] -0.5782141  0.5782141
#> [21,]  0.4188211 -0.4188211
#> [22,]  0.4139615 -0.4139615
#> [23,] -0.5803464  0.5803464
#> [24,] -0.5781476  0.5781476
#> [25,] -0.5803692  0.5803692
#> [26,] -0.5759939  0.5759939
#> [27,]  0.4157767 -0.4157767
#> [28,]  0.4134685 -0.4134685
#> [29,]  0.4126509 -0.4126509
#> [30,]  0.4104345 -0.4104345
#> [31,]  0.4136787 -0.4136787
#> [32,] -0.5877900  0.5877900
#> [33,]  0.4098373 -0.4098373
#> [34,] -0.5876918  0.5876918
#> [35,] -0.5745447  0.5745447
#> [36,]  0.4123358 -0.4123358
#> [37,]  0.4141876 -0.4141876
#> [38,]  0.4077072 -0.4077072
#> [39,] -0.5861747  0.5861747
#> [40,]  0.4188980 -0.4188980
#> 
#> $L
#> [1] 0.6941178 0.7032514 0.7366819 0.7030979 0.6831336 0.7112495
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319