Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.3532301  0.01438414 -0.08819465  0.04894256  0.064185498 -0.01943121
#>  [2,] -0.4041865 -0.12514388 -0.06601284 -0.13731131 -0.058984539  0.03052555
#>  [3,] -0.3671641 -0.08177556 -0.13202919 -0.16170960 -0.159165232 -0.05459650
#>  [4,] -0.3464172 -0.18637333 -0.02036820 -0.17873861 -0.085673900 -0.05143608
#>  [5,] -0.3608474 -0.11887049 -0.14264467 -0.07216517  0.048901044 -0.03493226
#>  [6,] -0.2405666 -0.14337686 -0.04843819 -0.05784010 -0.144350512 -0.11294228
#>  [7,] -0.2928234 -0.13020251 -0.11431507 -0.11911525 -0.208504300 -0.02667292
#>  [8,] -0.1703548 -0.10627882 -0.05246675 -0.06146903 -0.031249302 -0.06223994
#>  [9,] -0.2815346 -0.14064601 -0.01244680 -0.08727223 -0.032245937 -0.04860221
#> [10,] -0.3198301 -0.01882842 -0.14981447 -0.06390551 -0.003958486 -0.13397414
#>                V15         V16         V17         V18         V19         V2
#>  [1,] -0.007251747  0.03483573  0.18551346  0.23329210  0.28402029 -0.4065505
#>  [2,]  0.055713311 -0.04212525  0.12004554  0.04949962  0.04312145 -0.2541023
#>  [3,] -0.099343251  0.11742442  0.13120497 -0.03743008  0.08672751 -0.3564324
#>  [4,] -0.022903264 -0.07514474  0.02308772 -0.06032327  0.00607727 -0.2488237
#>  [5,]  0.114582353  0.01738068  0.03067819  0.15849083  0.09201800 -0.3697699
#>  [6,]  0.005548716 -0.08110449 -0.06312441  0.03822505  0.13222990 -0.2675283
#>  [7,] -0.032596743  0.04694634 -0.13420774  0.03494080  0.05272031 -0.3779660
#>  [8,] -0.044403050  0.03059568  0.05666411  0.05983137 -0.05145357 -0.2778391
#>  [9,] -0.110143918 -0.03620217 -0.14560791 -0.11086730 -0.10229621 -0.1966537
#> [10,] -0.092423436 -0.05582780 -0.06080340 -0.04325646 -0.02541163 -0.2228485
#>                V20          V21         V22          V23        V24         V25
#>  [1,]  0.208472026  0.240999407  0.22426962  0.203155796 0.28414369  0.24699735
#>  [2,]  0.165563573  0.184889569  0.13703165  0.179663789 0.02696046  0.23037472
#>  [3,]  0.089852444  0.171548722  0.10273683  0.099131449 0.01310979 -0.05306348
#>  [4,]  0.105725955  0.005451113  0.10547067 -0.003912128 0.09281840  0.11689352
#>  [5,]  0.182893662  0.106142847  0.07848538  0.181508552 0.23518371  0.15418711
#>  [6,]  0.118242316  0.050261046  0.21037947  0.138332284 0.09718117  0.09461086
#>  [7,]  0.026697762  0.145772241 -0.03432834 -0.004430922 0.06044969  0.04929778
#>  [8,] -0.009892704  0.074188316  0.05212839  0.068324288 0.04066437  0.11448998
#>  [9,]  0.058123412 -0.008126691  0.13089763  0.055512200 0.06414350  0.15017850
#> [10,]  0.008172545  0.002917518  0.03318319  0.149301276 0.12548288  0.05049255
#>               V26         V27        V28         V29         V3          V30
#>  [1,]  0.19071818  0.28281371 0.14676526 0.143796008 -0.3961959  0.069630481
#>  [2,]  0.11123625  0.22179083 0.09372664 0.139963681 -0.3344780  0.009829351
#>  [3,]  0.03329342  0.10533802 0.17850647 0.057119927 -0.3555521  0.080110605
#>  [4,] -0.01163025 -0.01032301 0.14302084 0.077863078 -0.3306234  0.043143177
#>  [5,]  0.11022570  0.25374336 0.07696029 0.101022020 -0.1967988  0.024254987
#>  [6,]  0.23131601  0.14247845 0.06912866 0.088238887 -0.2046541  0.085950066
#>  [7,]  0.15473653 -0.04253828 0.02413972 0.005354013 -0.3787319  0.062477545
#>  [8,]  0.12173243  0.11312355 0.19032044 0.038802916 -0.1993464 -0.029382674
#>  [9,]  0.20806262  0.09737270 0.17029770 0.154551011 -0.2047043 -0.015263043
#> [10,]  0.24623429  0.24213637 0.21503759 0.253167528 -0.3022686  0.205969440
#>               V31          V32          V33          V34         V35
#>  [1,]  0.02831466  0.057211371  0.100443120  0.004960238 -0.01210321
#>  [2,]  0.08245516 -0.173245976 -0.036463151 -0.173125711 -0.10617510
#>  [3,] -0.01306018 -0.015165881  0.024134879  0.001404573  0.03681058
#>  [4,]  0.09972579 -0.083028280  0.007238779 -0.103066789 -0.01441339
#>  [5,]  0.02634929 -0.005092153 -0.111084959 -0.120222983 -0.03092190
#>  [6,] -0.05186770  0.064949545 -0.003652499 -0.049242006 -0.09368307
#>  [7,] -0.16746978 -0.149691611 -0.099206232  0.010792947  0.01019055
#>  [8,]  0.02945290  0.035261280 -0.039935804 -0.109824286 -0.09692889
#>  [9,]  0.07881127  0.056134556 -0.078033122  0.096875068 -0.04370963
#> [10,]  0.13352398  0.040389774  0.060818704  0.075346311  0.12151533
#>               V36         V37          V38          V39         V4         V40
#>  [1,] -0.05678999 -0.04419212 -0.068063396  0.010986105 -0.2994085 -0.10736746
#>  [2,] -0.02083487 -0.05652980 -0.033372163 -0.227143024 -0.3646449 -0.20926617
#>  [3,]  0.02415883 -0.02526468  0.024342068 -0.053947383 -0.2626371 -0.01168069
#>  [4,] -0.06819391 -0.11469714 -0.128919598 -0.006501102 -0.1873756 -0.16115083
#>  [5,] -0.06483401 -0.07432405 -0.122574284 -0.081138036 -0.3447703 -0.03913504
#>  [6,] -0.09332110 -0.12786290 -0.180710847 -0.126235065 -0.1920330 -0.14404841
#>  [7,] -0.12077551 -0.11462255 -0.007917225 -0.151617768 -0.2378078 -0.18879752
#>  [8,]  0.02000834 -0.06326216 -0.025261107 -0.019292217 -0.2822316  0.02464271
#>  [9,]  0.02182694 -0.04926125 -0.067998901 -0.089653185 -0.3369322  0.03112061
#> [10,] -0.02815943  0.05347206 -0.054548601 -0.022618031 -0.2316235 -0.03389232
#>                V41          V42         V43         V44         V45         V46
#>  [1,] -0.089907272 -0.132991501 -0.13086066 -0.04921700 -0.03871274 -0.08985255
#>  [2,] -0.193614878 -0.219698779 -0.05350640 -0.16927875 -0.16910852 -0.11001910
#>  [3,] -0.075454072 -0.003198265 -0.07553499 -0.18629282 -0.05357397 -0.16186347
#>  [4,] -0.059403260 -0.061389910 -0.04586287 -0.10072341 -0.11430301 -0.17147452
#>  [5,] -0.008925962 -0.003603539 -0.08045196 -0.06146144 -0.10804253 -0.13921469
#>  [6,] -0.167866444 -0.059395405 -0.07558752 -0.21472671 -0.08928546 -0.18724730
#>  [7,] -0.158654256 -0.223654221 -0.20390348 -0.14231040 -0.21494607 -0.26962678
#>  [8,] -0.003540152 -0.072785502 -0.16327201 -0.09163838 -0.04431607 -0.12789309
#>  [9,] -0.081511864 -0.168939047 -0.15489777 -0.08876375 -0.22546637 -0.13925984
#> [10,] -0.137843266  0.002589692 -0.09588282 -0.09280342 -0.17106524 -0.18649127
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.1021694 -0.2257312 -0.3876952 -0.3066492 -0.3991051 -0.3055785
#>  [2,] -0.1818647 -0.2291147 -0.2301381 -0.1898541 -0.3797568 -0.2749154
#>  [3,] -0.1967368 -0.3026657 -0.2635881 -0.2204279 -0.2855199 -0.3654063
#>  [4,] -0.1453756 -0.1664674 -0.2420859 -0.3157591 -0.3178170 -0.3381840
#>  [5,] -0.2647457 -0.2816960 -0.3237505 -0.1881143 -0.3313021 -0.3819186
#>  [6,] -0.2258186 -0.2658346 -0.3418761 -0.3146528 -0.2333856 -0.3258429
#>  [7,] -0.3266529 -0.1436759 -0.2651656 -0.1806980 -0.4019284 -0.3236196
#>  [8,] -0.2133205 -0.2297188 -0.1660936 -0.2883050 -0.2508176 -0.2800542
#>  [9,] -0.1271872 -0.2822116 -0.2346828 -0.2310713 -0.2355257 -0.3839169
#> [10,] -0.1774880 -0.1089275 -0.3245338 -0.2158968 -0.2591235 -0.3741141
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3462035 -0.4459409 -0.3481309 -0.4519932 -0.3007053 -0.3685335
#>  [2,] -0.4467460 -0.3422896 -0.3305464 -0.2753198 -0.4407519 -0.4871238
#>  [3,] -0.2444000 -0.3800089 -0.4114044 -0.2292839 -0.3473345 -0.3450031
#>  [4,] -0.3498389 -0.2812446 -0.3190671 -0.3595049 -0.3686939 -0.2561057
#>  [5,] -0.2896677 -0.4056282 -0.2684728 -0.3185279 -0.3738576 -0.3390284
#>  [6,] -0.3089809 -0.2270673 -0.2745692 -0.3668627 -0.2619049 -0.2848306
#>  [7,] -0.2382921 -0.3780578 -0.2627644 -0.2693256 -0.3927070 -0.3095659
#>  [8,] -0.2289940 -0.2230679 -0.3603863 -0.2376553 -0.2023323 -0.2562012
#>  [9,] -0.3749083 -0.3631706 -0.3616970 -0.3760241 -0.3287107 -0.3967303
#> [10,] -0.4139626 -0.3354701 -0.4077655 -0.3970440 -0.2546965 -0.2891803
#>              V58        V59          V6        V60          V7          V8
#>  [1,] -0.4498352 -0.4072112 -0.15861782 -0.4159779 -0.22055784 -0.09696763
#>  [2,] -0.3482573 -0.3746228 -0.28346001 -0.4086255 -0.32553569 -0.23863116
#>  [3,] -0.2423366 -0.3306876 -0.24447731 -0.3061926 -0.19827049 -0.23105192
#>  [4,] -0.4013748 -0.2483916 -0.24788600 -0.3911124 -0.27907121 -0.17911652
#>  [5,] -0.4136325 -0.4402630 -0.27514260 -0.2532731 -0.27364304 -0.19817669
#>  [6,] -0.3322033 -0.2343097 -0.16331941 -0.3509919 -0.10224317 -0.15684003
#>  [7,] -0.2637610 -0.4091126 -0.22312228 -0.4021043 -0.15809507 -0.10689224
#>  [8,] -0.3344242 -0.3608959 -0.07904127 -0.2494103 -0.09657292 -0.07249093
#>  [9,] -0.3086287 -0.2960545 -0.23724910 -0.3748124 -0.13900911 -0.15130480
#> [10,] -0.3041375 -0.4253360 -0.22030785 -0.3319873 -0.13041861 -0.20509883
#>                V9
#>  [1,] -0.09549553
#>  [2,] -0.23180636
#>  [3,] -0.16085798
#>  [4,] -0.11339157
#>  [5,] -0.04150833
#>  [6,] -0.19923732
#>  [7,] -0.15649330
#>  [8,] -0.02922390
#>  [9,] -0.13108315
#> [10,] -0.07443244
#> 
#> $W[[2]]
#>             [,1]        [,2]         [,3]        [,4]       [,5]        [,6]
#> [1,]  0.03941991  0.02008464 -0.003684914 -0.10993364 0.09163082 -0.08332103
#> [2,] -0.01609759 -0.07072172 -0.017703028 -0.04349257 0.06434753 -0.03751737
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.04954629 -0.07210118 -0.14274277  0.02726361
#> [2,]  0.02418174  0.01434246 -0.01609017 -0.04841833
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  2.813863e-05  0.0002687571  0.0002410266  0.0002616190  0.0002962616
#>  [2,] -1.220545e-04 -0.0007046465 -0.0009531914 -0.0008399387 -0.0006319213
#>  [3,] -3.394836e-05 -0.0001435660 -0.0002363597 -0.0001812613 -0.0000741296
#>  [4,]  1.015011e-04  0.0007577648  0.0009189968  0.0009054954  0.0008265215
#>  [5,]  3.052865e-05  0.0003251262  0.0003062069  0.0003301715  0.0003833342
#>  [6,]  1.031716e-04  0.0007759259  0.0009181616  0.0009014113  0.0008481363
#>  [7,]  5.227633e-05  0.0003747741  0.0004871675  0.0004980153  0.0004349710
#>  [8,]  1.844861e-04  0.0013360552  0.0015957504  0.0015212219  0.0013772837
#>  [9,]  2.113524e-04  0.0014711345  0.0017743778  0.0016785282  0.0015142395
#> [10,] -1.227379e-04 -0.0006745121 -0.0008927969 -0.0007609758 -0.0005839223
#>                 V14           V15           V16          V17          V18
#>  [1,]  3.596710e-04  0.0003584379  3.897725e-04 0.0003413089 0.0003104864
#>  [2,] -3.450806e-04 -0.0001593041 -1.251271e-05 0.0001949996 0.0002752728
#>  [3,]  6.274972e-05  0.0001393864  2.257073e-04 0.0003189679 0.0003554355
#>  [4,]  7.049372e-04  0.0005916434  5.447276e-04 0.0004704670 0.0004706555
#>  [5,]  4.561628e-04  0.0004577546  5.005535e-04 0.0004795299 0.0004628476
#>  [6,]  7.572100e-04  0.0006558155  6.276736e-04 0.0005450311 0.0005272204
#>  [7,]  3.405569e-04  0.0002694059  2.360278e-04 0.0002297481 0.0002628641
#>  [8,]  1.174930e-03  0.0009845418  8.958416e-04 0.0006997924 0.0006445464
#>  [9,]  1.280203e-03  0.0010817742  9.738873e-04 0.0007485147 0.0007084872
#> [10,] -3.539154e-04 -0.0002323797 -9.184872e-05 0.0001221088 0.0001812651
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0004113812  4.311172e-05  4.828947e-04  4.768985e-04  0.0004820156
#>  [2,] -0.0002770089 -1.645100e-04 -9.743797e-04 -1.088410e-03 -0.0005508378
#>  [3,]  0.0001871821 -4.362591e-05 -5.390281e-05 -7.740245e-05  0.0001412558
#>  [4,]  0.0008958092  1.408568e-04  1.352719e-03  1.453640e-03  0.0012174427
#>  [5,]  0.0005755429  4.961935e-05  6.372942e-04  6.421990e-04  0.0006793645
#>  [6,]  0.0009570591  1.465062e-04  1.402363e-03  1.496851e-03  0.0012751560
#>  [7,]  0.0004753419  6.657951e-05  7.180676e-04  7.900699e-04  0.0006505595
#>  [8,]  0.0014077928  2.586369e-04  2.259140e-03  2.417614e-03  0.0019272637
#>  [9,]  0.0015657127  2.889960e-04  2.558723e-03  2.754556e-03  0.0022059267
#> [10,] -0.0003269064 -1.617903e-04 -9.993470e-04 -1.119453e-03 -0.0006467343
#>                V23          V24          V25          V26          V27
#>  [1,] 5.467937e-04 0.0005500213 0.0005423515 0.0005731318 0.0006215045
#>  [2,] 2.627412e-04 0.0006153880 0.0009029018 0.0006698733 0.0005156832
#>  [3,] 4.706845e-04 0.0006284710 0.0007591235 0.0006806151 0.0006315684
#>  [4,] 8.397962e-04 0.0006874904 0.0005374084 0.0006653255 0.0007814187
#>  [5,] 7.608228e-04 0.0007957917 0.0008098535 0.0008292870 0.0008670755
#>  [6,] 9.233593e-04 0.0007825687 0.0006471285 0.0007816372 0.0008956533
#>  [7,] 4.284985e-04 0.0003332534 0.0002380783 0.0002923005 0.0003561070
#>  [8,] 1.192070e-03 0.0008942626 0.0006167947 0.0008731821 0.0010734118
#>  [9,] 1.367511e-03 0.0010328821 0.0007205452 0.0010159742 0.0012112815
#> [10,] 6.690053e-05 0.0003715706 0.0006619451 0.0004550651 0.0003531771
#>                V28          V29            V3           V30          V31
#>  [1,] 0.0006684468 6.132800e-04  5.486431e-05  5.981685e-04 5.545680e-04
#>  [2,] 0.0004892375 3.864848e-04 -1.049105e-04  2.116182e-04 8.732281e-04
#>  [3,] 0.0006064438 4.970225e-04 -1.515412e-05  3.880047e-04 6.067892e-04
#>  [4,] 0.0008069623 7.545392e-04  1.209456e-04  7.344512e-04 2.543824e-04
#>  [5,] 0.0008958805 7.884156e-04  6.306921e-05  7.303559e-04 6.786534e-04
#>  [6,] 0.0009217829 8.379414e-04  1.303848e-04  8.207240e-04 3.642812e-04
#>  [7,] 0.0003614241 3.650170e-04  5.437092e-05  3.438992e-04 6.737278e-05
#>  [8,] 0.0011393707 1.063834e-03  2.160577e-04  1.107500e-03 2.578457e-04
#>  [9,] 0.0012948369 1.264159e-03  2.401630e-04  1.317161e-03 3.153634e-04
#> [10,] 0.0002595944 9.365083e-05 -1.109028e-04 -7.240028e-05 5.846549e-04
#>                V32           V33           V34           V35           V36
#>  [1,] 0.0005326261  5.437363e-04  5.615404e-04  0.0005099672  0.0004284579
#>  [2,] 0.0009420628  1.057116e-03  1.403753e-03  0.0018139260  0.0017134135
#>  [3,] 0.0006320264  6.821660e-04  7.941491e-04  0.0009166118  0.0008507606
#>  [4,] 0.0002002638  1.319080e-04 -1.091995e-04 -0.0004031187 -0.0003907817
#>  [5,] 0.0006766205  6.877881e-04  6.966072e-04  0.0006367639  0.0005463932
#>  [6,] 0.0003231249  2.700192e-04  5.510643e-05 -0.0002421487 -0.0002662703
#>  [7,] 0.0000215147 -2.497745e-05 -1.976004e-04 -0.0003538313 -0.0003048526
#>  [8,] 0.0001780895  7.333705e-05 -3.105074e-04 -0.0008183245 -0.0008219142
#>  [9,] 0.0002061939  1.149673e-04 -3.173432e-04 -0.0008871532 -0.0008792137
#> [10,] 0.0006579098  6.836835e-04  9.117944e-04  0.0012351483  0.0011664398
#>                 V37          V38          V39            V4          V40
#>  [1,]  4.267372e-04 4.303141e-04 0.0003804599  7.025870e-05 3.925003e-04
#>  [2,]  1.144089e-03 6.059526e-04 0.0002725344 -1.276000e-04 5.404019e-04
#>  [3,]  6.372217e-04 4.386386e-04 0.0003000603 -1.351167e-05 3.937623e-04
#>  [4,] -8.592729e-05 2.183862e-04 0.0003817436  1.560209e-04 2.340986e-04
#>  [5,]  5.372056e-04 5.396611e-04 0.0004915774  8.398346e-05 4.946344e-04
#>  [6,]  2.214757e-05 3.178137e-04 0.0004608228  1.686994e-04 3.110979e-04
#>  [7,] -1.296299e-04 3.338005e-05 0.0001419878  7.184670e-05 7.383038e-05
#>  [8,] -2.604796e-04 3.074094e-04 0.0005884713  2.763170e-04 2.955989e-04
#>  [9,] -3.009014e-04 3.180717e-04 0.0006448089  3.062223e-04 3.034246e-04
#> [10,]  8.002432e-04 3.834389e-04 0.0001079112 -1.265565e-04 3.616196e-04
#>                V41           V42           V43           V44           V45
#>  [1,] 4.409026e-04  4.008293e-04  0.0003359279  0.0003260689  3.560382e-04
#>  [2,] 4.354685e-04  1.440201e-05 -0.0000782867 -0.0001082288 -5.390121e-04
#>  [3,] 3.587491e-04  1.799605e-04  0.0001361448  0.0001180686 -3.999868e-05
#>  [4,] 2.806571e-04  4.831914e-04  0.0005090737  0.0004967856  7.159680e-04
#>  [5,] 5.342809e-04  4.849593e-04  0.0004304164  0.0004192853  4.407160e-04
#>  [6,] 3.793979e-04  5.595692e-04  0.0005737476  0.0005624609  7.887440e-04
#>  [7,] 6.139812e-05  1.882889e-04  0.0002072070  0.0001882561  2.832836e-04
#>  [8,] 4.154847e-04  7.885592e-04  0.0008337148  0.0008342464  1.280402e-03
#>  [9,] 4.190028e-04  8.516367e-04  0.0009186322  0.0008914513  1.366490e-03
#> [10,] 3.076140e-04 -5.937568e-05 -0.0001683345 -0.0001451889 -4.685174e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  2.666146e-04  1.998394e-04  1.572986e-04  7.917614e-05  9.035513e-05
#>  [2,] -5.266195e-04 -2.995145e-04 -2.279494e-04 -1.482518e-04 -1.323867e-04
#>  [3,] -8.539286e-05 -3.281842e-05 -2.179884e-05 -1.957349e-05 -2.433753e-06
#>  [4,]  5.909829e-04  3.840596e-04  2.947146e-04  1.754333e-04  2.043341e-04
#>  [5,]  3.187256e-04  2.293430e-04  1.787565e-04  9.272154e-05  1.158644e-04
#>  [6,]  6.368614e-04  4.143509e-04  3.200451e-04  1.868913e-04  2.183654e-04
#>  [7,]  2.524485e-04  1.715642e-04  1.300760e-04  8.124793e-05  9.826648e-05
#>  [8,]  1.062488e-03  6.730872e-04  5.195750e-04  3.063893e-04  3.474141e-04
#>  [9,]  1.143836e-03  7.159514e-04  5.573132e-04  3.324900e-04  3.820095e-04
#> [10,] -4.553853e-04 -2.476598e-04 -2.004849e-04 -1.306014e-04 -1.320960e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  2.778960e-05  2.428992e-05  1.678728e-05  1.608273e-05  1.719285e-05
#>  [2,] -1.850025e-05 -4.055613e-05 -3.733804e-05  1.099353e-06 -8.564859e-06
#>  [3,]  7.181489e-06 -4.346785e-06 -5.445810e-06  8.122254e-06  3.553468e-06
#>  [4,]  4.375846e-05  5.162883e-05  4.288307e-05  1.912688e-05  2.261208e-05
#>  [5,]  3.420277e-05  2.869636e-05  2.060702e-05  1.979063e-05  1.894218e-05
#>  [6,]  4.997526e-05  5.528296e-05  4.555826e-05  2.284946e-05  2.546106e-05
#>  [7,]  1.746005e-05  2.327275e-05  1.981685e-05  6.947899e-06  9.570058e-06
#>  [8,]  7.524546e-05  8.985298e-05  7.544739e-05  3.181041e-05  3.892923e-05
#>  [9,]  8.290046e-05  9.898737e-05  8.429895e-05  3.582211e-05  4.394317e-05
#> [10,] -2.104712e-05 -3.942311e-05 -3.729171e-05 -4.945998e-06 -1.330431e-05
#>                 V55           V56          V57           V58           V59
#>  [1,]  7.477172e-06  1.259892e-05 1.112081e-05  7.000324e-06  1.405567e-05
#>  [2,] -1.522873e-05 -8.757699e-06 7.743453e-06 -2.027268e-05 -3.839227e-07
#>  [3,] -6.221253e-07  2.453785e-06 8.132778e-06 -4.495250e-06  5.376094e-06
#>  [4,]  1.983122e-05  1.955913e-05 9.479221e-06  2.066755e-05  1.582348e-05
#>  [5,]  1.051000e-05  1.521460e-05 1.370898e-05  8.156449e-06  1.641740e-05
#>  [6,]  2.148511e-05  2.222369e-05 1.193212e-05  2.124551e-05  1.878698e-05
#>  [7,]  9.415787e-06  7.954911e-06 3.164006e-06  1.096160e-05  5.630561e-06
#>  [8,]  3.432717e-05  3.427949e-05 1.517602e-05  3.619030e-05  2.739288e-05
#>  [9,]  3.951646e-05  3.799621e-05 1.725782e-05  4.242523e-05  2.993797e-05
#> [10,] -1.599809e-05 -1.113544e-05 1.945700e-06 -2.312031e-05 -6.202583e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  9.922386e-05  9.057204e-06  1.171740e-04  1.384961e-04  0.0002346608
#>  [2,] -1.369480e-04 -6.420399e-07 -8.721107e-05 -2.947478e-04 -0.0006449685
#>  [3,]  1.860851e-06  3.459444e-06  3.246757e-05 -4.563401e-05 -0.0001417062
#>  [4,]  2.378233e-04  1.156737e-05  2.285770e-04  3.549026e-04  0.0006566447
#>  [5,]  1.322262e-04  1.093111e-05  1.555079e-04  1.691305e-04  0.0002728304
#>  [6,]  2.491628e-04  1.333848e-05  2.463517e-04  3.662314e-04  0.0006720815
#>  [7,]  1.175354e-04  4.659447e-06  1.099352e-04  1.776422e-04  0.0003242657
#>  [8,]  3.977393e-04  1.966560e-05  3.802251e-04  6.213118e-04  0.0011714518
#>  [9,]  4.404139e-04  2.235628e-05  4.184741e-04  6.969067e-04  0.0012986151
#> [10,] -1.488941e-04 -5.410710e-06 -1.051546e-04 -3.079567e-04 -0.0006246102
#> 
#> $vW[[2]]
#>              [,1]         [,2]        [,3]        [,4]         [,5]
#> [1,] -0.001522718  0.006511595  0.01434343  0.01262168  0.003803783
#> [2,]  0.001522718 -0.006511595 -0.01434343 -0.01262168 -0.003803783
#>              [,6]        [,7]         [,8]        [,9]       [,10]
#> [1,]  0.006985809  0.01666864  0.008385578  0.01614973  0.01587346
#> [2,] -0.006985809 -0.01666864 -0.008385578 -0.01614973 -0.01587346
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.7132573 0.4097724 0.2975330 0.3801329 0.6166325 0.3228918 0.2688542
#>  [8] 0.2998004 0.5713205 0.4688692
#> 
#> $B[[2]]
#> [1]  0.14295277 -0.09895056
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0009852028 0.0010395583 0.0009897733 0.0009251852 0.0012984311
#>  [6] 0.0011255635 0.0003867564 0.0012732412 0.0014712914 0.0005778249
#> 
#> $vB[[2]]
#> [1]  0.0008027047 -0.0008027047
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#>  [2,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [3,] 0.1088 0.5761 0.4733 0.2362 0.1023 0.2904 0.4713 0.4659 0.1415 0.0849
#>  [4,] 0.0374 0.2597 0.3483 0.3999 0.4574 0.5950 0.7924 0.8272 0.8087 0.8977
#>  [5,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#>  [6,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#>  [7,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#>  [8,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#>  [9,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [10,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [11,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [12,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#> [13,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#> [14,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [15,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [16,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [17,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [18,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [19,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [20,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [21,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [22,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [23,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [24,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [25,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [26,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [27,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [28,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [29,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [30,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [31,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [32,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [33,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [34,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [35,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [36,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [37,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [38,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#> [39,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [40,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#>  [2,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [3,] 0.3257 0.1278 0.9007 0.9312 0.4856 0.1346 0.1604 0.2737 0.5609 0.3654
#>  [4,] 0.9828 0.0586 0.8982 0.8890 0.9367 0.9122 0.7936 0.6718 0.6318 0.4865
#>  [5,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#>  [6,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#>  [7,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#>  [8,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#>  [9,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [10,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [11,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [12,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#> [13,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#> [14,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [15,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [16,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [17,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [18,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [19,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [20,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [21,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [22,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [23,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [24,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [25,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [26,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [27,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [28,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [29,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [30,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [31,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [32,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [33,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [34,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [35,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [36,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [37,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [38,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#> [39,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [40,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#>  [2,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [3,] 0.6139 0.5470 0.0926 0.8474 0.5638 0.5443 0.5086 0.6253 0.8497 0.8406
#>  [4,] 0.3388 0.4832 0.0628 0.3822 0.3075 0.1267 0.0743 0.1510 0.1906 0.1817
#>  [5,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#>  [6,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#>  [7,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#>  [8,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#>  [9,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [10,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [11,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [12,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#> [13,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#> [14,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [15,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [16,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [17,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [18,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [19,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [20,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [21,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [22,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [23,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [24,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [25,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [26,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [27,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [28,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [29,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [30,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [31,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [32,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [33,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [34,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [35,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [36,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [37,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [38,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#> [39,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [40,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#>  [2,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [3,] 0.8420 0.9136 0.7713 0.1234 0.4882 0.3724 0.4469 0.4586 0.4491 0.5616
#>  [4,] 0.1709 0.0946 0.2829 0.0534 0.3006 0.1602 0.1483 0.2875 0.2047 0.1064
#>  [5,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#>  [6,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#>  [7,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#>  [8,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#>  [9,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [10,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [11,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [12,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#> [13,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#> [14,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [15,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [16,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [17,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [18,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [19,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [20,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [21,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [22,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [23,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [24,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [25,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [26,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [27,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [28,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [29,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [30,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [31,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [32,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [33,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [34,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [35,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [36,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [37,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [38,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#> [39,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [40,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#>  [2,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [3,] 0.4305 0.0945 0.0794 0.0274 0.1276 0.0154 0.0140 0.0455 0.0213 0.0082
#>  [4,] 0.1395 0.1065 0.0527 0.0395 0.0255 0.0183 0.0353 0.0118 0.0063 0.0237
#>  [5,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#>  [6,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#>  [7,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#>  [8,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#>  [9,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [10,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [11,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [12,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#> [13,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#> [14,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [15,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [16,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [17,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [18,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [19,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [20,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [21,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [22,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [23,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [24,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [25,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [26,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [27,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [28,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [29,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [30,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [31,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [32,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [33,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [34,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [35,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [36,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [37,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [38,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#> [39,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [40,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#>  [2,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [3,] 0.0124 0.0167 0.0103 0.0205 0.0178 0.1731 0.0187 0.1948 0.4262 0.6828
#>  [4,] 0.0032 0.0087 0.0124 0.0113 0.0098 0.1422 0.0126 0.2072 0.2734 0.3070
#>  [5,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#>  [6,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#>  [7,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#>  [8,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#>  [9,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [10,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [11,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [12,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#> [13,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> [14,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [15,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [16,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [17,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [18,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [19,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [20,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [21,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [22,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [23,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [24,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [25,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [26,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [27,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [28,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [29,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [30,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [31,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [32,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [33,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [34,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [35,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [36,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [37,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [38,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#> [39,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [40,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9402905 0.7474238 0.6379187 0.5664251 0.8389256 0.6853894 0.4356091
#>  [2,] 0.9598100 0.8042676 0.7154805 0.6062391 0.8709714 0.7465235 0.5137750
#>  [3,] 0.7543798 0.3475821 0.4775731 0.3386540 0.5268386 0.3433845 0.2175506
#>  [4,] 0.9277061 0.6890708 0.5867668 0.5040188 0.8087929 0.6227242 0.4009459
#>  [5,] 0.8544210 0.6800228 0.5998992 0.5847054 0.7671186 0.6496937 0.4945728
#>  [6,] 0.9101289 0.6008763 0.5870353 0.4446654 0.7586935 0.5208566 0.3334569
#>  [7,] 0.9268469 0.6796851 0.5821588 0.4845765 0.8020377 0.5968106 0.3823155
#>  [8,] 0.9481585 0.7883029 0.6809803 0.5979622 0.8560471 0.7253550 0.5201122
#>  [9,] 0.8779566 0.5914483 0.5630048 0.4507315 0.7439200 0.5169800 0.3389424
#> [10,] 0.9132695 0.7347278 0.6825578 0.5954872 0.8019869 0.6858515 0.4834997
#> [11,] 0.8845153 0.6556875 0.5944320 0.5627791 0.7619026 0.6338885 0.4051414
#> [12,] 0.8604049 0.5270870 0.5226730 0.4566556 0.6892088 0.5132321 0.3176372
#> [13,] 0.8050278 0.5218373 0.5341142 0.4803737 0.6740828 0.5093950 0.3618685
#> [14,] 0.8628586 0.5202090 0.5668112 0.4800330 0.7140785 0.5309313 0.2909666
#> [15,] 0.9351402 0.7246890 0.6467155 0.5382693 0.8232017 0.6542255 0.4159030
#> [16,] 0.8991119 0.6281088 0.6493732 0.5616804 0.7552706 0.6494972 0.4065588
#> [17,] 0.8249891 0.4847982 0.5232474 0.4195180 0.6686131 0.4732630 0.2415049
#> [18,] 0.9050921 0.6183788 0.5497677 0.4513047 0.7543418 0.5537412 0.3416600
#> [19,] 0.9385512 0.7372476 0.6223941 0.5780073 0.8247194 0.7062451 0.4492267
#> [20,] 0.7583048 0.4320868 0.5921888 0.4666206 0.6113826 0.4284389 0.2883489
#> [21,] 0.8886159 0.6836617 0.6248429 0.5876586 0.7837104 0.6617181 0.4701180
#> [22,] 0.9205127 0.7007917 0.6379298 0.5571390 0.8125390 0.6618278 0.4742005
#> [23,] 0.9564395 0.8209143 0.7254706 0.6325118 0.8743828 0.7606229 0.5526010
#> [24,] 0.9136455 0.7122561 0.6462836 0.5820136 0.8042619 0.6637667 0.4784066
#> [25,] 0.8882737 0.6588311 0.6184485 0.5861443 0.7675684 0.6470558 0.4154164
#> [26,] 0.9291046 0.7009004 0.6010425 0.5342518 0.8221683 0.6698522 0.4091097
#> [27,] 0.8786070 0.6035559 0.5622956 0.4657840 0.7343723 0.5153684 0.3148838
#> [28,] 0.9374738 0.7339562 0.6794536 0.5747168 0.8202727 0.6883398 0.4664677
#> [29,] 0.9128894 0.7023190 0.6635739 0.5428682 0.7876735 0.6416695 0.4460972
#> [30,] 0.7245658 0.4254303 0.5816082 0.4593132 0.5844475 0.4203260 0.3008016
#> [31,] 0.9441669 0.7482361 0.6641668 0.5865720 0.8446897 0.7002078 0.4295040
#> [32,] 0.8593836 0.6070063 0.6118646 0.5448078 0.7289649 0.6000106 0.4122279
#> [33,] 0.7825786 0.4070591 0.5202372 0.4156517 0.6200900 0.4358029 0.2573870
#> [34,] 0.7508235 0.3679994 0.4975932 0.3532430 0.5100423 0.3454607 0.2616427
#> [35,] 0.8904381 0.6719613 0.6034428 0.5727221 0.7752924 0.6453933 0.4131140
#> [36,] 0.8635037 0.5193255 0.5453753 0.4615920 0.6832927 0.4947954 0.3297595
#> [37,] 0.9241361 0.7443459 0.6428873 0.6098192 0.8226725 0.7179182 0.4597281
#> [38,] 0.8731268 0.5268203 0.5262748 0.4495671 0.7131164 0.5172028 0.2415922
#> [39,] 0.9507785 0.7680420 0.6543905 0.5696224 0.8522484 0.7055992 0.4353697
#> [40,] 0.8898772 0.6603424 0.5995639 0.5603282 0.7659770 0.6346223 0.4067069
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6633850 0.6755190 0.7659508
#>  [2,] 0.6906328 0.6997058 0.7873977
#>  [3,] 0.4474883 0.4668371 0.6095797
#>  [4,] 0.5950540 0.5700781 0.6442319
#>  [5,] 0.6457262 0.7162015 0.7783281
#>  [6,] 0.5972258 0.5683532 0.6963484
#>  [7,] 0.5974551 0.5627959 0.6415874
#>  [8,] 0.6822951 0.6853112 0.7454541
#>  [9,] 0.5754620 0.5279847 0.6361199
#> [10,] 0.6560741 0.7229300 0.7936716
#> [11,] 0.6277571 0.7036278 0.7955937
#> [12,] 0.5225947 0.5846184 0.6899470
#> [13,] 0.5512193 0.5659846 0.6478278
#> [14,] 0.5647567 0.6198021 0.7316781
#> [15,] 0.6250964 0.6161603 0.7129539
#> [16,] 0.5996508 0.7174874 0.8226406
#> [17,] 0.5360762 0.5743557 0.6906235
#> [18,] 0.5523388 0.5231522 0.6148111
#> [19,] 0.6534700 0.7112818 0.7956506
#> [20,] 0.5343344 0.6178882 0.7688082
#> [21,] 0.6439293 0.7157650 0.7762263
#> [22,] 0.6191674 0.6203436 0.7079352
#> [23,] 0.7042800 0.7279438 0.8091471
#> [24,] 0.6618532 0.6774036 0.7632117
#> [25,] 0.6172065 0.7240984 0.8322021
#> [26,] 0.6340589 0.6487735 0.7376093
#> [27,] 0.5799859 0.5735936 0.6903256
#> [28,] 0.6225639 0.6261877 0.7274788
#> [29,] 0.6095557 0.6127033 0.6937116
#> [30,] 0.5161090 0.6057438 0.7341017
#> [31,] 0.6815031 0.7098369 0.8066360
#> [32,] 0.5870514 0.6997828 0.8050360
#> [33,] 0.5197243 0.5460931 0.6496644
#> [34,] 0.4470696 0.4622048 0.5700177
#> [35,] 0.6380085 0.7069616 0.7983244
#> [36,] 0.5462141 0.6121848 0.7173610
#> [37,] 0.6668289 0.7398252 0.8249731
#> [38,] 0.5652297 0.5805874 0.7189244
#> [39,] 0.6679057 0.6793518 0.7779751
#> [40,] 0.6269921 0.6924409 0.7908203
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5833514 0.4166486
#>  [2,] 0.5840164 0.4159836
#>  [3,] 0.5779956 0.4220044
#>  [4,] 0.5827787 0.4172213
#>  [5,] 0.5787481 0.4212519
#>  [6,] 0.5833155 0.4166845
#>  [7,] 0.5830612 0.4169388
#>  [8,] 0.5825407 0.4174593
#>  [9,] 0.5819289 0.4180711
#> [10,] 0.5818313 0.4181687
#> [11,] 0.5806848 0.4193152
#> [12,] 0.5803767 0.4196233
#> [13,] 0.5777217 0.4222783
#> [14,] 0.5803631 0.4196369
#> [15,] 0.5840121 0.4159879
#> [16,] 0.5814308 0.4185692
#> [17,] 0.5801419 0.4198581
#> [18,] 0.5825352 0.4174648
#> [19,] 0.5825027 0.4174973
#> [20,] 0.5786755 0.4213245
#> [21,] 0.5797815 0.4202185
#> [22,] 0.5820938 0.4179062
#> [23,] 0.5834622 0.4165378
#> [24,] 0.5814179 0.4185821
#> [25,] 0.5813388 0.4186612
#> [26,] 0.5825820 0.4174180
#> [27,] 0.5825456 0.4174544
#> [28,] 0.5836577 0.4163423
#> [29,] 0.5826321 0.4173679
#> [30,] 0.5774293 0.4225707
#> [31,] 0.5834547 0.4165453
#> [32,] 0.5804334 0.4195666
#> [33,] 0.5772156 0.4227844
#> [34,] 0.5770124 0.4229876
#> [35,] 0.5809329 0.4190671
#> [36,] 0.5800631 0.4199369
#> [37,] 0.5820660 0.4179340
#> [38,] 0.5816883 0.4183117
#> [39,] 0.5842526 0.4157474
#> [40,] 0.5811075 0.4188925
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]        [,2]         [,3]        [,4]       [,5]        [,6]
#>  [1,] 2.7566977  1.08491936  0.566342142  0.26728019 1.65025593  0.77865167
#>  [2,] 3.1731179  1.41318354  0.922152505  0.43153056 1.90957476  1.08015595
#>  [3,] 1.1221095 -0.62968429 -0.089767663 -0.66929803 0.10745755 -0.64824770
#>  [4,] 2.5519750  0.79577872  0.350615507  0.01607569 1.44218602  0.50112738
#>  [5,] 1.7697052  0.75387675  0.405045053  0.34211997 1.19211229  0.61769306
#>  [6,] 2.3152095  0.40911760  0.351723027 -0.22224891 1.14553007  0.08347490
#>  [7,] 2.5392343  0.75232511  0.331641821 -0.06171358 1.39907884  0.39219325
#>  [8,] 2.9063315  1.31472622  0.758280473  0.39698138 1.78283884  0.97118167
#>  [9,] 1.9732207  0.36995598  0.253365848 -0.19771551 1.06644377  0.06794621
#> [10,] 2.3542250  1.01874365  0.765551256  0.38669635 1.39875930  0.78079511
#> [11,] 2.0359014  0.64413444  0.382317742  0.25244879 1.16313908  0.54893521
#> [12,] 1.8186571  0.10845438  0.090754320 -0.17381405 0.79642294  0.05294078
#> [13,] 1.4180201  0.08740465  0.136669095 -0.07854551 0.72670963  0.03758453
#> [14,] 1.8392387  0.08088024  0.268852791 -0.07991049 0.91527557  0.12388320
#> [15,] 2.6684688  0.96784143  0.604632801  0.15337709 1.53819151  0.63766515
#> [16,] 2.1873954  0.52411209  0.616285223  0.24798468 1.12692292  0.61682967
#> [17,] 1.5505221 -0.06082589  0.093056513 -0.32475222 0.70191881 -0.10705018
#> [18,] 2.2551295  0.48267253  0.199732167 -0.19540071 1.12190446  0.21579818
#> [19,] 2.7261330  1.03171164  0.499722587  0.31459834 1.54865488  0.87721660
#> [20,] 1.1434079 -0.27334211  0.373021169 -0.13371629 0.45312749 -0.28822329
#> [21,] 2.0766802  0.77065078  0.510155317  0.35429443 1.28742123  0.67096029
#> [22,] 2.4493330  0.85107062  0.566390330  0.22955886 1.46659288  0.67145004
#> [23,] 3.0890678  1.52255426  0.971762426  0.54300755 1.94027889  1.15609764
#> [24,] 2.3589811  0.90636664  0.602742998  0.33104489 1.41314724  0.68012557
#> [25,] 2.0732279  0.65808972  0.482968194  0.34804881 1.19463162  0.60612236
#> [26,] 2.5730151  0.85158934  0.409810824  0.13722230 1.53110773  0.70751681
#> [27,] 1.9793049  0.42030347  0.250483864 -0.13707824 1.01692040  0.06149316
#> [28,] 2.7076023  1.01478842  0.751261879  0.30112211 1.51819609  0.79236909
#> [29,] 2.3494364  0.85836516  0.679261745  0.17189483 1.31095843  0.58261747
#> [30,] 0.9672238 -0.30052009  0.329378629 -0.16310772 0.34105817 -0.32143523
#> [31,] 2.8279364  1.08922694  0.681918765  0.34981222 1.69354436  0.84828757
#> [32,] 1.8101798  0.43474574  0.455156902  0.17971335 0.98937732  0.40550921
#> [33,] 1.2807568 -0.37613622  0.080993117 -0.34064964 0.48993027 -0.25821377
#> [34,] 1.1030094 -0.54080908 -0.009627272 -0.60481455 0.04017471 -0.63905294
#> [35,] 2.0952235  0.71706911  0.419830760  0.29296606 1.23844046  0.59885077
#> [36,] 1.8447007  0.07734069  0.182002056 -0.15393519 0.76894538 -0.02081923
#> [37,] 2.4999182  1.06868064  0.587918223  0.44655237 1.53456022  0.93415830
#> [38,] 1.9288927  0.10738417  0.105195927 -0.20241978 0.91056822  0.06883840
#> [39,] 2.9609513  1.19728813  0.638394761  0.28031056 1.75234563  0.87410506
#> [40,] 2.0894878  0.66482032  0.403648530  0.24249423 1.18573286  0.55209847
#>              [,7]        [,8]        [,9]     [,10]
#>  [1,] -0.25900197  0.67841587  0.73325441 1.1855869
#>  [2,]  0.05511389  0.80307920  0.84589738 1.3093103
#>  [3,] -1.27999758 -0.21082423 -0.13284647 0.4455458
#>  [4,] -0.40152538  0.38489825  0.28216968 0.5937797
#>  [5,] -0.02170956  0.60030558  0.92569718 1.2559495
#>  [6,] -0.69259126  0.39391933  0.27513551 0.8299692
#>  [7,] -0.47973167  0.39487249  0.25251708 0.5822604
#>  [8,]  0.08049211  0.76433928  0.77828863 1.0745126
#>  [9,] -0.66801101  0.30417151  0.11205603 0.5585624
#> [10,] -0.06602517  0.64584757  0.95904232 1.3472006
#> [11,] -0.38408773  0.52260673  0.86463359 1.3589787
#> [12,] -0.76465180  0.09044028  0.34176172 0.7998713
#> [13,] -0.56726338  0.20559851  0.26548702 0.6095045
#> [14,] -0.89069398  0.26049002  0.48870843 1.0031533
#> [15,] -0.33961488  0.51123693  0.47328205 0.9097740
#> [16,] -0.37820958  0.40401042  0.93203203 1.5343412
#> [17,] -1.14444668  0.14455592  0.29964482 0.8030360
#> [18,] -0.65590532  0.21012510  0.09267502 0.4675810
#> [19,] -0.20379547  0.63432706  0.90161737 1.3593286
#> [20,] -0.90341645  0.13755403  0.48059448 1.2015940
#> [21,] -0.11967069  0.59245941  0.92355038 1.2438089
#> [22,] -0.10328993  0.48601591  0.49100686 0.8853770
#> [23,]  0.21118547  0.86776280  0.98421524 1.4444780
#> [24,] -0.08642755  0.67156348  0.74186547 1.1703692
#> [25,] -0.34161853  0.47770818  0.96488281 1.6013153
#> [26,] -0.36764687  0.54966948  0.61365247 1.0335799
#> [27,] -0.77738458  0.32271547  0.29652827 0.8016421
#> [28,] -0.13433082  0.50044510  0.51589657 0.9818687
#> [29,] -0.21645246  0.44544501  0.45868992 0.8175292
#> [30,] -0.84348349  0.06445843  0.42945603 1.0155337
#> [31,] -0.28387526  0.76068817  0.89459181 1.4282982
#> [32,] -0.35476295  0.35178909  0.84626385 1.4180719
#> [33,] -1.05959428  0.07893809  0.18489743 0.6175642
#> [34,] -1.03744812 -0.21251788 -0.15146992 0.2819234
#> [35,] -0.35110712  0.56673081  0.88067291 1.3758546
#> [36,] -0.70927292  0.18538563  0.45650516 0.9314088
#> [37,] -0.16143743  0.69387743  1.04506015 1.5504113
#> [38,] -1.14397027  0.26241433  0.32518505 0.9391326
#> [39,] -0.25997554  0.69872792  0.75079447 1.2539050
#> [40,] -0.37759593  0.51933435  0.81155567 1.3298773
#> 
#> $pre[[3]]
#>             [,1]       [,2]
#>  [1,] 0.06962609 -0.2669203
#>  [2,] 0.06077882 -0.2785043
#>  [3,] 0.10932276 -0.2052276
#>  [4,] 0.08862363 -0.2455670
#>  [5,] 0.05283746 -0.2647990
#>  [6,] 0.09893240 -0.2374663
#>  [7,] 0.09278507 -0.2425676
#>  [8,] 0.06087782 -0.2723343
#>  [9,] 0.09922063 -0.2314762
#> [10,] 0.05643679 -0.2738587
#> [11,] 0.06253658 -0.2630486
#> [12,] 0.09012355 -0.2341966
#> [13,] 0.08181248 -0.2316151
#> [14,] 0.08478814 -0.2394759
#> [15,] 0.08085395 -0.2584116
#> [16,] 0.06238077 -0.2662690
#> [17,] 0.09533466 -0.2280213
#> [18,] 0.10003333 -0.2331561
#> [19,] 0.06198706 -0.2710689
#> [20,] 0.08194924 -0.2353894
#> [21,] 0.05575727 -0.2661193
#> [22,] 0.07385124 -0.2575234
#> [23,] 0.05311037 -0.2838919
#> [24,] 0.06292944 -0.2656673
#> [25,] 0.05960265 -0.2686691
#> [26,] 0.07519243 -0.2581895
#> [27,] 0.09437404 -0.2388583
#> [28,] 0.07296343 -0.2648435
#> [29,] 0.07632079 -0.2572672
#> [30,] 0.07993277 -0.2322963
#> [31,] 0.06398972 -0.2729818
#> [32,] 0.06492141 -0.2596315
#> [33,] 0.09256345 -0.2187899
#> [34,] 0.10422016 -0.2063005
#> [35,] 0.06150242 -0.2651019
#> [36,] 0.08642830 -0.2366040
#> [37,] 0.05412095 -0.2771392
#> [38,] 0.09484924 -0.2348588
#> [39,] 0.06991842 -0.2703371
#> [40,] 0.06469716 -0.2626244
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4166486 -0.4166486
#>  [2,] -0.5840164  0.5840164
#>  [3,]  0.4220044 -0.4220044
#>  [4,]  0.4172213 -0.4172213
#>  [5,] -0.5787481  0.5787481
#>  [6,] -0.5833155  0.5833155
#>  [7,]  0.4169388 -0.4169388
#>  [8,] -0.5825407  0.5825407
#>  [9,] -0.5819289  0.5819289
#> [10,] -0.5818313  0.5818313
#> [11,]  0.4193152 -0.4193152
#> [12,]  0.4196233 -0.4196233
#> [13,]  0.4222783 -0.4222783
#> [14,] -0.5803631  0.5803631
#> [15,]  0.4159879 -0.4159879
#> [16,] -0.5814308  0.5814308
#> [17,]  0.4198581 -0.4198581
#> [18,]  0.4174648 -0.4174648
#> [19,]  0.4174973 -0.4174973
#> [20,] -0.5786755  0.5786755
#> [21,] -0.5797815  0.5797815
#> [22,] -0.5820938  0.5820938
#> [23,] -0.5834622  0.5834622
#> [24,] -0.5814179  0.5814179
#> [25,] -0.5813388  0.5813388
#> [26,]  0.4174180 -0.4174180
#> [27,] -0.5825456  0.5825456
#> [28,]  0.4163423 -0.4163423
#> [29,]  0.4173679 -0.4173679
#> [30,] -0.5774293  0.5774293
#> [31,]  0.4165453 -0.4165453
#> [32,] -0.5804334  0.5804334
#> [33,] -0.5772156  0.5772156
#> [34,]  0.4229876 -0.4229876
#> [35,]  0.4190671 -0.4190671
#> [36,] -0.5800631  0.5800631
#> [37,]  0.4179340 -0.4179340
#> [38,]  0.4183117 -0.4183117
#> [39,]  0.4157474 -0.4157474
#> [40,]  0.4188925 -0.4188925
#> 
#> $L
#> [1] 0.7020142 0.7153706 0.7018744 0.6972819 0.7225016 0.6976759
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4492754