Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13         V14
#>  [1,] -0.2429149 -0.11214018 -0.08992406  0.01923929 -0.06694688 -0.16311393
#>  [2,] -0.4156527 -0.07581676 -0.19378976 -0.13347698 -0.01354976  0.01129335
#>  [3,] -0.2370813 -0.05815388 -0.19275970 -0.18468507 -0.02348788 -0.17790413
#>  [4,] -0.1855218 -0.19434907 -0.10924570 -0.05359667 -0.09279706 -0.11786665
#>  [5,] -0.2903594 -0.21467189 -0.10409809 -0.21239008 -0.09605634 -0.06545937
#>  [6,] -0.2278420 -0.20989334 -0.10882488 -0.20942930 -0.21294888 -0.19788000
#>  [7,] -0.2848810 -0.16663517 -0.05584185 -0.04550702 -0.09177337 -0.09367809
#>  [8,] -0.3047863 -0.06409738 -0.11441161 -0.18723659 -0.20057853 -0.07350957
#>  [9,] -0.4486373 -0.17747932 -0.06184328 -0.05823360 -0.03353608 -0.01339152
#> [10,] -0.2508794 -0.19884266 -0.18786680 -0.10094225 -0.03418791 -0.05040109
#>                 V15         V16          V17         V18          V19
#>  [1,] -0.2114928375 -0.10669212 -0.082267121 -0.03344100 -0.105608178
#>  [2,] -0.0372621468 -0.05051126  0.130042685  0.09202658  0.097499295
#>  [3,] -0.1347709004  0.02696430 -0.039954810 -0.01713748 -0.095546742
#>  [4,] -0.1911483692 -0.05585863 -0.211446195 -0.01629259 -0.109745424
#>  [5,] -0.1010791639 -0.13595261 -0.002029058 -0.12550222  0.003749235
#>  [6,] -0.0800304938  0.02468187 -0.017096391  0.03977380 -0.054496688
#>  [7,]  0.0587265749 -0.08580906  0.125402108  0.06480934 -0.008075203
#>  [8,] -0.1226829138 -0.05502976 -0.071021821  0.01481386  0.061095568
#>  [9,] -0.0009659796  0.11172444  0.130958061  0.13452748  0.074022119
#> [10,] -0.0049048094 -0.12567859 -0.089096049 -0.07358313 -0.027036954
#>               V2          V20         V21          V22          V23        V24
#>  [1,] -0.3863134 -0.149922583 -0.12582048 -0.005863788  0.097811183 0.03162703
#>  [2,] -0.3368866  0.110844094  0.14815775  0.124982116  0.130123245 0.18584072
#>  [3,] -0.2288506 -0.002232048  0.06784279  0.029613310 -0.011348781 0.15712898
#>  [4,] -0.1797769  0.030954435 -0.08912873  0.106638991  0.139996820 0.19127710
#>  [5,] -0.3246446 -0.006941467  0.02052156  0.080142681  0.008218842 0.12190736
#>  [6,] -0.1962902 -0.051544658  0.01591050  0.008642180  0.020365268 0.13632220
#>  [7,] -0.2766456  0.068970468  0.08268306  0.216843057  0.189785863 0.15072475
#>  [8,] -0.3308344  0.078414791  0.03046119 -0.016696863  0.170388426 0.09684320
#>  [9,] -0.3635272  0.244219075  0.25938440  0.261120022  0.242910697 0.27778907
#> [10,] -0.3066496  0.057419302  0.14410309  0.074725800  0.084867269 0.18841044
#>                V25        V26          V27         V28          V29         V3
#>  [1,]  0.131949076 0.16198679  0.022352896  0.06676796  0.082884077 -0.3550023
#>  [2,]  0.168777165 0.12505580  0.139146709  0.21015013  0.225063509 -0.2748648
#>  [3,] -0.003379565 0.04518710 -0.013339691  0.13662228 -0.080073282 -0.3346850
#>  [4,]  0.118932904 0.12914953  0.222282672  0.08983739  0.049885515 -0.3467506
#>  [5,]  0.106249911 0.02331467  0.069889189 -0.08152722  0.041365222 -0.2275342
#>  [6,]  0.091397573 0.01998224  0.002465775 -0.07310080  0.004483512 -0.3487351
#>  [7,]  0.152696814 0.03931167  0.184789649  0.14597100  0.063323732 -0.2387520
#>  [8,]  0.203536036 0.17267775  0.245252727  0.26621296  0.278749372 -0.3438223
#>  [9,]  0.230884207 0.24385469  0.198459405  0.21255067  0.136920857 -0.2223517
#> [10,]  0.115760578 0.19410585  0.265378338  0.25389336  0.110131775 -0.2484438
#>               V30         V31         V32          V33         V34          V35
#>  [1,]  0.15051501  0.07163141  0.04876754 -0.006039216 -0.05743766 -0.035613466
#>  [2,]  0.19623451  0.13684374  0.04491509  0.064403849  0.04058989  0.005156439
#>  [3,] -0.04455820 -0.14113964 -0.22163182 -0.225559773 -0.21858137 -0.067690358
#>  [4,] -0.05991379 -0.11496722 -0.08254240  0.038850047 -0.03396537 -0.065302486
#>  [5,] -0.08933837 -0.12044305 -0.10827789 -0.142803903 -0.13372360 -0.095587407
#>  [6,]  0.06111715 -0.10738525  0.04047037 -0.033155878  0.03743208 -0.062386005
#>  [7,]  0.02736420 -0.07736127 -0.03629868 -0.052939862 -0.08542470 -0.012700661
#>  [8,]  0.08251772  0.11623266  0.06692886  0.045093296  0.13598749 -0.046272735
#>  [9,]  0.16434364  0.17278648  0.18575844  0.096899309  0.14878738  0.123871230
#> [10,]  0.13021110  0.01785053 -0.01381618  0.096230910  0.06196759 -0.032796931
#>                V36          V37         V38          V39         V4         V40
#>  [1,]  0.021763657  0.030563664 -0.06795036 -0.012565690 -0.3299458 -0.03616156
#>  [2,]  0.025514919  0.050049617  0.01313951 -0.031922910 -0.3462165 -0.10432748
#>  [3,] -0.201895164 -0.089233072 -0.24020006 -0.251928696 -0.3370238 -0.08902815
#>  [4,] -0.136754874 -0.078296319  0.02616581  0.027967824 -0.2783928 -0.03531451
#>  [5,] -0.137098112 -0.219243111 -0.21149590 -0.127701347 -0.2336730 -0.20887597
#>  [6,] -0.040329122 -0.133748263 -0.12459979 -0.176996202 -0.1523702 -0.17465420
#>  [7,] -0.111407253 -0.002172032 -0.03805366 -0.032547839 -0.2693771 -0.16140641
#>  [8,]  0.086790271  0.028026071  0.05452440  0.026126069 -0.2538973 -0.10023604
#>  [9,]  0.007252826  0.003932374  0.02791605 -0.009840578 -0.2103567  0.10924033
#> [10,]  0.098660674  0.079075381  0.04922928  0.033331573 -0.3346306 -0.01603882
#>                 V41          V42         V43         V44         V45
#>  [1,] -0.0682414120  0.042100370 -0.03814574 -0.13877379 -0.19073925
#>  [2,] -0.0861482073 -0.169345037 -0.10404611 -0.05670140 -0.10456250
#>  [3,] -0.2382405057 -0.132295479 -0.07175257 -0.08830966 -0.10873468
#>  [4,] -0.0549252569 -0.042070998 -0.12997802 -0.16133393 -0.22643233
#>  [5,] -0.1976572750 -0.036316755 -0.04498540 -0.17947344 -0.08770328
#>  [6,] -0.0890835569 -0.095963690 -0.01549843 -0.14526800 -0.24018258
#>  [7,]  0.0002933957  0.004395538 -0.12521805 -0.13346631 -0.06126858
#>  [8,] -0.1089059443 -0.102336316 -0.12213047 -0.18890705 -0.11443655
#>  [9,] -0.0457891069 -0.055634425 -0.10786069 -0.07088019 -0.07173513
#> [10,] -0.0136219077 -0.133667688 -0.17601029 -0.09290222 -0.13370142
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.11217754 -0.1205536 -0.1916995 -0.2822214 -0.2061056 -0.3817320
#>  [2,] -0.14622314 -0.3298584 -0.2430287 -0.3343447 -0.2211861 -0.4828476
#>  [3,] -0.23716278 -0.2305731 -0.3014346 -0.3167275 -0.1599081 -0.2544674
#>  [4,] -0.24474358 -0.2352440 -0.3103440 -0.2006727 -0.1516090 -0.2273481
#>  [5,] -0.07556189 -0.1342883 -0.1885005 -0.2736483 -0.1752077 -0.2078186
#>  [6,] -0.23778507 -0.1515617 -0.2931283 -0.1729292 -0.1950083 -0.2163264
#>  [7,] -0.12854350 -0.1771578 -0.2035148 -0.3200772 -0.2384257 -0.3267083
#>  [8,] -0.16608456 -0.2455410 -0.1700914 -0.1561046 -0.3498299 -0.3246552
#>  [9,] -0.20329918 -0.1521138 -0.1643213 -0.2710444 -0.3008410 -0.3611355
#> [10,] -0.24286647 -0.2835453 -0.1634897 -0.3731976 -0.3257100 -0.3087080
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2487097 -0.3694520 -0.4132996 -0.3052236 -0.3192700 -0.3629962
#>  [2,] -0.4646107 -0.3124200 -0.4095373 -0.4999009 -0.3408813 -0.3556208
#>  [3,] -0.3580839 -0.3560574 -0.2102566 -0.3533055 -0.3685074 -0.3475314
#>  [4,] -0.3269124 -0.2765089 -0.2087624 -0.3874253 -0.3199858 -0.4164743
#>  [5,] -0.2982212 -0.3532480 -0.3884118 -0.3181983 -0.3669700 -0.2697232
#>  [6,] -0.3006172 -0.3376139 -0.3535525 -0.2055387 -0.2618764 -0.2403565
#>  [7,] -0.2659375 -0.2359521 -0.3071859 -0.3538494 -0.2826068 -0.2420643
#>  [8,] -0.2734778 -0.4130013 -0.2453720 -0.4117935 -0.3857087 -0.4158677
#>  [9,] -0.3269255 -0.3032828 -0.3634844 -0.3062951 -0.4309673 -0.4134612
#> [10,] -0.3864987 -0.4205708 -0.4162210 -0.3361413 -0.2974236 -0.2621880
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.3949806 -0.3563600 -0.3827918 -0.1774196 -0.2470583 -0.27463978
#>  [2,] -0.3538006 -0.4483150 -0.3407461 -0.2660733 -0.4192227 -0.19009609
#>  [3,] -0.2951410 -0.1982331 -0.3284074 -0.1735636 -0.2694571 -0.21216430
#>  [4,] -0.2850827 -0.3943248 -0.2893989 -0.1304574 -0.3593029 -0.13576396
#>  [5,] -0.2275626 -0.2474140 -0.2563739 -0.1401890 -0.3681318 -0.23395376
#>  [6,] -0.3572610 -0.3281333 -0.3474251 -0.1454034 -0.2346700 -0.09121615
#>  [7,] -0.3633611 -0.3364485 -0.2770369 -0.2744701 -0.3443799 -0.19499827
#>  [8,] -0.3473765 -0.2688553 -0.3080711 -0.2532944 -0.4078063 -0.10626142
#>  [9,] -0.3696311 -0.4569922 -0.3656997 -0.2492971 -0.4698519 -0.15320538
#> [10,] -0.2789101 -0.2922850 -0.4455672 -0.2152338 -0.3440428 -0.16728195
#>               V8          V9
#>  [1,] -0.1756934 -0.15747146
#>  [2,] -0.1766667 -0.11261797
#>  [3,] -0.2037710 -0.22383847
#>  [4,] -0.1803278 -0.07861597
#>  [5,] -0.2634106 -0.11351262
#>  [6,] -0.1915239 -0.15135157
#>  [7,] -0.2079846 -0.18267997
#>  [8,] -0.2442335 -0.24167333
#>  [9,] -0.1711165 -0.07458392
#> [10,] -0.2137908 -0.17669449
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.05474745 -0.06245884 -0.09410225 -0.05229806 -0.03447338 -0.04546469
#> [2,]  0.06288814  0.06997333  0.11739915  0.03376297  0.01348799  0.09080389
#>           [,7]       [,8]        [,9]       [,10]
#> [1,] 0.0301927 0.03738612 -0.04598584 -0.14253476
#> [2,] 0.1025611 0.13644762 -0.10498411  0.05672105
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12          V13
#>  [1,]  2.264889e-05  1.500821e-04  2.061083e-04  2.152737e-04 3.068250e-04
#>  [2,]  3.945055e-05  3.804682e-04  4.632638e-04  4.493045e-04 4.793021e-04
#>  [3,]  4.035742e-05  6.386702e-04  7.290558e-04  7.926188e-04 3.826819e-04
#>  [4,]  1.724177e-05  3.605809e-05  8.862732e-05  9.371607e-05 2.690001e-04
#>  [5,]  4.830213e-06 -8.421230e-05 -7.080368e-05 -7.967793e-05 9.934131e-05
#>  [6,]  1.708513e-05  1.467291e-04  1.784149e-04  1.886246e-04 1.914788e-04
#>  [7,]  2.684121e-05  1.655057e-04  2.502089e-04  2.522627e-04 4.239804e-04
#>  [8,]  5.726219e-06  6.033289e-05  1.130799e-04  1.242607e-04 2.733916e-04
#>  [9,] -3.252465e-05 -1.401539e-04 -1.252892e-04 -1.278230e-04 5.054527e-05
#> [10,]  8.102982e-05  7.384839e-04  8.606404e-04  8.581953e-04 7.411529e-04
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0003490848  0.0004413332  0.0006064547  0.0006844856  0.0007851007
#>  [2,]  0.0004009704  0.0004045369  0.0005670161  0.0006891321  0.0008019329
#>  [3,] -0.0001249332 -0.0005547679 -0.0008122578 -0.0007117295 -0.0006327966
#>  [4,]  0.0003896511  0.0005630494  0.0007963196  0.0008752195  0.0009744024
#>  [5,]  0.0002803842  0.0004712118  0.0006522119  0.0006733924  0.0007215264
#>  [6,]  0.0001835525  0.0001954925  0.0002369851  0.0002736786  0.0003328922
#>  [7,]  0.0004767810  0.0006127671  0.0008913984  0.0010183523  0.0011635222
#>  [8,]  0.0003439090  0.0004987952  0.0007582761  0.0008506726  0.0009479042
#>  [9,]  0.0001796984  0.0002950050  0.0004743481  0.0005020545  0.0005827569
#> [10,]  0.0005016555  0.0003591720  0.0004291171  0.0005897393  0.0007410077
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0008741037  2.465122e-05  0.0010035684  0.0011638255  0.0011555854
#>  [2,]  0.0009567511  4.202736e-05  0.0011477783  0.0013237454  0.0012968486
#>  [3,] -0.0004111114  4.887021e-05 -0.0004481940 -0.0005701339 -0.0005319249
#>  [4,]  0.0010271436  1.629734e-05  0.0011736526  0.0013821286  0.0013707908
#>  [5,]  0.0007189792  3.860326e-06  0.0008143254  0.0009596766  0.0009592626
#>  [6,]  0.0004058158  2.067256e-05  0.0004625725  0.0005159407  0.0005150220
#>  [7,]  0.0012694936  2.623597e-05  0.0014813697  0.0017442070  0.0017322051
#>  [8,]  0.0009865001  5.249675e-06  0.0011408094  0.0013590551  0.0013061439
#>  [9,]  0.0005473563 -2.749224e-05  0.0006301912  0.0007307829  0.0007237155
#> [10,]  0.0010592053  8.722813e-05  0.0012894374  0.0014380700  0.0014355075
#>                 V23          V24           V25           V26           V27
#>  [1,]  0.0012888333  0.001461004  0.0014465749  0.0014720933  0.0014564916
#>  [2,]  0.0013215908  0.001436709  0.0014065847  0.0014394448  0.0014311575
#>  [3,] -0.0012211152 -0.001775237 -0.0018372896 -0.0019508250 -0.0017972419
#>  [4,]  0.0016363480  0.001923627  0.0019209586  0.0019716389  0.0019252704
#>  [5,]  0.0012640378  0.001540498  0.0015461737  0.0015856836  0.0015274052
#>  [6,]  0.0005129816  0.000533053  0.0005234565  0.0005184564  0.0005231649
#>  [7,]  0.0020096532  0.002332987  0.0023171306  0.0023873135  0.0023429339
#>  [8,]  0.0015264730  0.001808694  0.0017938701  0.0018496909  0.0017959042
#>  [9,]  0.0010680864  0.001386270  0.0013949619  0.0014734638  0.0013724459
#> [10,]  0.0012368407  0.001138223  0.0010610345  0.0010384623  0.0011035460
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0014085537  0.0013210526  3.246408e-05  0.0011471597  0.0009440911
#>  [2,]  0.0013751113  0.0012585419  4.237472e-05  0.0011037364  0.0008164417
#>  [3,] -0.0017433907 -0.0017216544  1.863939e-05 -0.0015661260 -0.0015368814
#>  [4,]  0.0018685948  0.0017798543  2.972730e-05  0.0015550717  0.0013261538
#>  [5,]  0.0014833377  0.0014235414  1.820944e-05  0.0012508684  0.0010930901
#>  [6,]  0.0005156569  0.0004891087  2.282669e-05  0.0004259880  0.0003296908
#>  [7,]  0.0022714690  0.0021502401  4.021129e-05  0.0018674443  0.0015332529
#>  [8,]  0.0017113592  0.0015938122  1.960504e-05  0.0013951204  0.0011648984
#>  [9,]  0.0012842339  0.0012287112 -8.208258e-06  0.0010215546  0.0008654063
#> [10,]  0.0010828962  0.0009314989  7.447114e-05  0.0008111905  0.0004927222
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0008695905  0.0008767682  0.0007429332  6.166790e-04  5.745454e-04
#>  [2,]  0.0008243759  0.0008612459  0.0006073531  4.066903e-04  3.681286e-04
#>  [3,] -0.0012259050 -0.0013238407 -0.0015359627 -1.581671e-03 -1.482848e-03
#>  [4,]  0.0012009999  0.0012268433  0.0010983644  9.640210e-04  8.940008e-04
#>  [5,]  0.0009403646  0.0009753982  0.0009687694  8.957578e-04  8.233973e-04
#>  [6,]  0.0002943136  0.0002989026  0.0002514238  1.890866e-04  1.698516e-04
#>  [7,]  0.0014388245  0.0014802810  0.0012356732  1.029645e-03  9.622189e-04
#>  [8,]  0.0010772055  0.0011380731  0.0009304599  7.595104e-04  7.374991e-04
#>  [9,]  0.0007604629  0.0008568304  0.0007575710  6.726130e-04  7.164946e-04
#> [10,]  0.0005914278  0.0005719851  0.0002796862  4.769402e-05 -1.137955e-05
#>                 V37           V38           V39            V4           V40
#>  [1,]  0.0005057428  0.0004229498  4.235992e-04  4.123798e-05  0.0003880120
#>  [2,]  0.0004000189  0.0004612211  5.374156e-04  7.509691e-05  0.0004531089
#>  [3,] -0.0009001788 -0.0002589748 -3.445776e-07  8.125207e-05 -0.0001329894
#>  [4,]  0.0007112312  0.0005044656  4.545839e-04  2.819435e-05  0.0004366594
#>  [5,]  0.0005971696  0.0003535680  2.752349e-04  5.024023e-06  0.0002974077
#>  [6,]  0.0001808194  0.0001996418  2.179725e-04  3.139303e-05  0.0001911415
#>  [7,]  0.0008007485  0.0006287118  6.201587e-04  5.326103e-05  0.0005733077
#>  [8,]  0.0006263009  0.0004480399  4.252630e-04  2.740220e-05  0.0003980162
#>  [9,]  0.0005353041  0.0002587164  2.221462e-04 -2.498759e-05  0.0002488202
#> [10,]  0.0001813102  0.0005234408  7.232637e-04  1.419212e-04  0.0005649037
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0002992554  2.630386e-04  3.004634e-04  2.971255e-04  1.834887e-04
#>  [2,]  0.0003435146  3.648419e-04  4.417543e-04  4.585927e-04  4.099265e-04
#>  [3,] -0.0001208378  1.747456e-04  2.419474e-04  2.380215e-04  4.994868e-04
#>  [4,]  0.0003414253  2.564279e-04  2.948769e-04  2.884234e-04  1.037347e-04
#>  [5,]  0.0002476057  1.438722e-04  1.412576e-04  1.312002e-04 -2.314816e-05
#>  [6,]  0.0001616595  1.677559e-04  1.721945e-04  1.574293e-04  1.330283e-04
#>  [7,]  0.0004369691  3.730584e-04  4.462595e-04  4.479536e-04  2.446022e-04
#>  [8,]  0.0002702054  1.628417e-04  2.287080e-04  2.563943e-04  6.815527e-05
#>  [9,]  0.0001013779 -5.623473e-05 -3.793626e-05 -4.147134e-06 -2.326643e-04
#> [10,]  0.0004808260  6.474715e-04  7.288705e-04  7.041081e-04  7.913348e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.100204e-04  7.909065e-05  5.374422e-05  3.293719e-05  6.333621e-05
#>  [2,]  2.613832e-04  1.575442e-04  1.232199e-04  8.114332e-05  1.056254e-04
#>  [3,]  3.479062e-04  2.406381e-04  2.175263e-04  1.474446e-04  1.034781e-04
#>  [4,]  5.290752e-05  3.859926e-05  1.343141e-05  5.469359e-06  4.721760e-05
#>  [5,] -1.827522e-05 -7.630253e-06 -2.298905e-05 -1.898459e-05  1.494420e-05
#>  [6,]  9.396591e-05  7.312253e-05  5.547495e-05  3.296052e-05  4.669985e-05
#>  [7,]  1.351894e-04  7.840108e-05  4.729918e-05  3.051034e-05  8.135027e-05
#>  [8,] -4.566377e-06 -1.011958e-05 -1.654096e-05 -8.302522e-06  4.240245e-05
#>  [9,] -2.475282e-04 -1.820941e-04 -1.394381e-04 -7.896595e-05 -1.727692e-05
#> [10,]  5.696720e-04  3.685989e-04  2.930006e-04  1.907121e-04  1.922655e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  1.603801e-05  1.241136e-05  1.584574e-05  1.466706e-05  9.495945e-06
#>  [2,]  2.515852e-05  2.028280e-05  2.131944e-05  1.702446e-05  1.191245e-05
#>  [3,]  8.059120e-06  2.287942e-05 -1.974827e-06 -9.475766e-06  1.694641e-06
#>  [4,]  1.384162e-05  9.299877e-06  1.772942e-05  1.770925e-05  9.521677e-06
#>  [5,]  9.882130e-06  2.991165e-06  1.179329e-05  1.290324e-05  6.609796e-06
#>  [6,]  1.107104e-05  9.480434e-06  8.251800e-06  6.457174e-06  5.975471e-06
#>  [7,]  1.810966e-05  1.350000e-05  2.358194e-05  2.217165e-05  1.173694e-05
#>  [8,]  7.863180e-06  1.381955e-06  1.263638e-05  1.413913e-05  5.548737e-06
#>  [9,] -1.537456e-05 -2.385199e-05 -1.841625e-06  5.961768e-06 -3.909416e-06
#> [10,]  4.892937e-05  4.798413e-05  3.385116e-05  2.194468e-05  2.241560e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  1.281015e-05  1.026294e-05  1.204246e-05  6.091351e-06  3.430301e-06
#>  [2,]  1.557989e-05  1.024005e-05  1.314370e-05  9.847391e-06  4.699910e-06
#>  [3,] -9.960220e-06 -1.068036e-05 -1.533776e-05  5.128030e-06  3.824389e-06
#>  [4,]  1.559889e-05  1.284580e-05  1.616490e-05  5.366539e-06  2.771023e-06
#>  [5,]  1.160645e-05  1.032976e-05  1.258483e-05  2.892785e-06  1.816005e-06
#>  [6,]  5.809212e-06  4.603740e-06  4.669294e-06  4.091027e-06  3.043023e-06
#>  [7,]  1.963661e-05  1.529284e-05  1.995408e-05  8.327923e-06  3.193825e-06
#>  [8,]  1.250514e-05  1.091906e-05  1.518703e-05  2.937320e-06 -6.299940e-07
#>  [9,]  6.094219e-06  6.543318e-06  8.816522e-06 -1.867719e-06 -6.431087e-06
#> [10,]  1.948496e-05  1.103271e-05  1.203058e-05  1.871727e-05  1.311704e-05
#>                 V6           V60            V7           V8            V9
#>  [1,] 9.296539e-05  3.725624e-06  1.577780e-04 1.403330e-04  1.609649e-04
#>  [2,] 1.162916e-04  3.889738e-06  1.920893e-04 2.093742e-04  3.142412e-04
#>  [3,] 2.432537e-05  2.134215e-06 -7.270489e-05 5.136613e-05  3.201561e-04
#>  [4,] 9.264507e-05  3.326915e-06  1.811977e-04 1.377671e-04  1.031032e-04
#>  [5,] 6.033948e-05  2.882064e-06  1.295454e-04 8.059419e-05  1.424659e-05
#>  [6,] 5.721866e-05  3.201734e-06  7.443665e-05 7.915600e-05  1.201166e-04
#>  [7,] 1.196595e-04  3.509205e-06  2.297612e-04 2.014184e-04  2.036793e-04
#>  [8,] 7.272980e-05  1.331067e-07  1.746375e-04 1.352934e-04  1.217570e-04
#>  [9,] 8.414804e-06 -5.452110e-06  8.392508e-05 4.883388e-05 -3.049227e-05
#> [10,] 1.864580e-04  1.029109e-05  2.291603e-04 3.018624e-04  5.385654e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.07619833  0.09317145  0.04958443  0.07266927  0.05149339  0.06819405
#> [2,] -0.07619833 -0.09317145 -0.04958443 -0.07266927 -0.05149339 -0.06819405
#>             [,7]       [,8]        [,9]       [,10]
#> [1,]  0.07623739  0.0927419  0.09634367  0.09163418
#> [2,] -0.07623739 -0.0927419 -0.09634367 -0.09163418
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3877896 0.5925327 0.1963005 0.2463666 0.1257420 0.2546279 0.2402461
#>  [8] 0.3700968 0.8456218 0.4710569
#> 
#> $B[[2]]
#> [1]  0.03840116 -0.03021826
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0019947021  0.0019400496 -0.0026926065  0.0026889367  0.0021580662
#>  [6]  0.0007274094  0.0032242803  0.0025358065  0.0019436816  0.0014020359
#> 
#> $vB[[2]]
#> [1]  0.1014141 -0.1014141
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#>  [2,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>  [3,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#>  [4,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#>  [5,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#>  [6,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#>  [7,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#>  [8,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#>  [9,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [10,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [11,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [12,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [13,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [14,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [15,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [16,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [17,] 0.0187 0.2684 0.3108 0.2933 0.2275 0.0994 0.1801 0.2200 0.2732 0.2862
#> [18,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [19,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [20,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [21,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [22,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#> [23,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [24,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [25,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [26,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [27,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [28,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [29,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [30,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [31,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [32,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [33,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [34,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [35,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [36,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [37,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [38,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [39,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [40,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#>  [2,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>  [3,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#>  [4,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#>  [5,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#>  [6,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#>  [7,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#>  [8,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#>  [9,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [10,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [11,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [12,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [13,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [14,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [15,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [16,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [17,] 0.2034 0.0346 0.1740 0.4130 0.6879 0.8120 0.8453 0.8919 0.9300 0.9987
#> [18,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [19,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [20,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [21,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [22,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#> [23,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [24,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [25,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [26,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [27,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [28,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [29,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [30,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [31,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [32,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [33,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [34,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [35,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [36,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [37,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [38,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [39,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [40,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#>  [2,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>  [3,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#>  [4,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#>  [5,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#>  [6,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#>  [7,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#>  [8,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#>  [9,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [10,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [11,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [12,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [13,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [14,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [15,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [16,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [17,] 1.0000 0.8104 0.0168 0.6199 0.6041 0.5547 0.4160 0.1472 0.0849 0.0608
#> [18,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [19,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [20,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [21,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [22,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#> [23,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [24,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [25,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [26,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [27,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [28,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [29,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [30,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [31,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [32,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [33,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [34,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [35,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [36,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [37,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [38,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [39,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [40,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#>  [2,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>  [3,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#>  [4,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#>  [5,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#>  [6,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#>  [7,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#>  [8,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#>  [9,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [10,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [11,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [12,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [13,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [14,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [15,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [16,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [17,] 0.0969 0.1411 0.1676 0.0177 0.1200 0.1201 0.1036 0.1977 0.1339 0.0902
#> [18,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [19,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [20,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [21,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [22,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#> [23,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [24,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [25,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [26,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [27,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [28,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [29,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [30,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [31,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [32,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [33,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [34,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [35,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [36,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [37,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [38,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [39,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [40,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#>  [2,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>  [3,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#>  [4,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#>  [5,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#>  [6,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#>  [7,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#>  [8,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#>  [9,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [10,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [11,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [12,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [13,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [14,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [15,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [16,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [17,] 0.1085 0.1521 0.1363 0.0858 0.0393 0.0290 0.0203 0.0116 0.0098 0.0199
#> [18,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [19,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [20,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [21,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [22,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#> [23,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [24,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [25,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [26,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [27,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [28,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [29,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [30,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [31,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [32,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [33,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [34,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [35,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [36,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [37,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [38,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [39,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [40,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#>  [2,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#>  [3,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#>  [4,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#>  [5,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#>  [6,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#>  [7,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#>  [8,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#>  [9,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [10,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [11,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [12,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [13,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [14,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [15,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [16,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [17,] 0.0033 0.0101 0.0065 0.0115 0.0193 0.1630 0.0157 0.2028 0.1694 0.2328
#> [18,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [19,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [20,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [21,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [22,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> [23,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [24,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [25,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [26,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [27,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [28,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [29,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [30,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [31,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [32,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [33,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [34,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [35,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [36,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [37,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [38,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [39,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [40,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5440332 0.8033742 0.3726312 0.5713735 0.3813084 0.4444972 0.6551847
#>  [2,] 0.6243113 0.8641188 0.2646643 0.5818163 0.3034912 0.4334779 0.6608439
#>  [3,] 0.6048366 0.8540111 0.3512512 0.5938206 0.3768143 0.4472784 0.6970974
#>  [4,] 0.5792743 0.7949442 0.1648442 0.4493716 0.2010761 0.3099851 0.5299345
#>  [5,] 0.4861445 0.8258609 0.2791923 0.4500833 0.2892151 0.3881987 0.6245857
#>  [6,] 0.6532899 0.8686549 0.2351277 0.5558735 0.2992635 0.4456175 0.6386589
#>  [7,] 0.5470605 0.8728893 0.3939650 0.5868126 0.4224499 0.5014643 0.7392560
#>  [8,] 0.6596980 0.8655888 0.3160255 0.6110090 0.3563419 0.4840556 0.6785919
#>  [9,] 0.4009037 0.7721137 0.2518488 0.4575012 0.2854679 0.3393446 0.6393097
#> [10,] 0.4972682 0.8478309 0.3234535 0.5328784 0.3397430 0.4016264 0.7134450
#> [11,] 0.6084163 0.8189471 0.3775010 0.6149313 0.4050533 0.4707405 0.6834211
#> [12,] 0.6527331 0.8412756 0.3574678 0.5878542 0.3883654 0.4934362 0.6521121
#> [13,] 0.5486601 0.7735124 0.2053606 0.4903297 0.2496435 0.3610528 0.5927665
#> [14,] 0.5057662 0.8672154 0.3280024 0.5350146 0.3440070 0.4048591 0.7244478
#> [15,] 0.5130369 0.8650309 0.3395399 0.5572115 0.3549524 0.4202852 0.7218042
#> [16,] 0.5854652 0.8224619 0.3037883 0.5663859 0.3389479 0.4309436 0.6183033
#> [17,] 0.6038750 0.8313170 0.3370002 0.5902305 0.3641197 0.4366720 0.6737179
#> [18,] 0.3334723 0.7334250 0.2521076 0.3913422 0.2463164 0.2974369 0.5732316
#> [19,] 0.6496893 0.8538107 0.3421307 0.6170704 0.3776396 0.4644028 0.6835261
#> [20,] 0.5719772 0.8352038 0.3832974 0.5473004 0.3850851 0.4786186 0.6730418
#> [21,] 0.4789761 0.7929145 0.1333905 0.4190394 0.1913443 0.2978474 0.5503833
#> [22,] 0.5610735 0.8741816 0.3207288 0.5761130 0.3385750 0.4164577 0.7258991
#> [23,] 0.4539980 0.6820261 0.1607708 0.3870197 0.1879781 0.2898189 0.4750050
#> [24,] 0.4746600 0.8211334 0.2055475 0.5021354 0.2380822 0.3156063 0.6180810
#> [25,] 0.4672390 0.7151662 0.1782396 0.4237927 0.2057439 0.2750282 0.5470297
#> [26,] 0.5689846 0.7902155 0.1531116 0.4364725 0.1890477 0.2989254 0.5296870
#> [27,] 0.5470410 0.8943762 0.4242444 0.5992212 0.4268049 0.5047199 0.7697612
#> [28,] 0.4496164 0.8451662 0.2757660 0.4850021 0.2955655 0.3608422 0.6918534
#> [29,] 0.6171822 0.8433164 0.3691480 0.6025152 0.3960413 0.4568111 0.6903362
#> [30,] 0.3517232 0.7922939 0.2584590 0.4124416 0.2774315 0.3189157 0.6259050
#> [31,] 0.3626361 0.7956287 0.2691567 0.3931648 0.2610934 0.3225841 0.6301680
#> [32,] 0.6302739 0.7870127 0.4186291 0.6154701 0.4407165 0.4939928 0.6543126
#> [33,] 0.5115388 0.8225263 0.2762883 0.5192871 0.3048881 0.3394229 0.6544996
#> [34,] 0.4879720 0.8599266 0.4075910 0.5672426 0.4104576 0.4710780 0.7378769
#> [35,] 0.6168072 0.8684561 0.2853297 0.5506709 0.3150452 0.4431508 0.6613989
#> [36,] 0.5546335 0.8404682 0.3311823 0.5220591 0.3654539 0.4465554 0.6772674
#> [37,] 0.4026944 0.6174259 0.1238325 0.2648208 0.1792935 0.2331170 0.4214864
#> [38,] 0.4970341 0.8894772 0.3829669 0.5622898 0.3855520 0.4757444 0.7584367
#> [39,] 0.5636484 0.7735459 0.1566694 0.4297650 0.1958341 0.3218707 0.4992416
#> [40,] 0.6475015 0.8736583 0.4325340 0.6455529 0.4555903 0.5206365 0.7400423
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7408236 0.9127088 0.7452191
#>  [2,] 0.8276458 0.9510054 0.8263171
#>  [3,] 0.8036133 0.9429319 0.8049493
#>  [4,] 0.7576899 0.9180308 0.7390004
#>  [5,] 0.7297815 0.9274454 0.7248989
#>  [6,] 0.8385269 0.9552784 0.8176301
#>  [7,] 0.7774174 0.9596244 0.7883373
#>  [8,] 0.8322776 0.9503146 0.8233420
#>  [9,] 0.6352785 0.9275190 0.6706724
#> [10,] 0.7482358 0.9534047 0.7620016
#> [11,] 0.7763484 0.9221430 0.7794886
#> [12,] 0.8077470 0.9259291 0.7941908
#> [13,] 0.6938298 0.9273452 0.7210721
#> [14,] 0.7693404 0.9601361 0.7689607
#> [15,] 0.7684341 0.9563295 0.7685251
#> [16,] 0.7772970 0.9192558 0.7701763
#> [17,] 0.7819519 0.9333062 0.7701023
#> [18,] 0.5313237 0.9045199 0.5757759
#> [19,] 0.8193908 0.9396012 0.8021395
#> [20,] 0.7654206 0.9283337 0.7454545
#> [21,] 0.6910382 0.9466354 0.6903791
#> [22,] 0.8031779 0.9605415 0.8011320
#> [23,] 0.6089583 0.8621611 0.6363227
#> [24,] 0.7390653 0.9460669 0.7499337
#> [25,] 0.6488804 0.9020776 0.6675117
#> [26,] 0.7500068 0.9227113 0.7326586
#> [27,] 0.8209394 0.9646438 0.8135189
#> [28,] 0.7216356 0.9541339 0.7259381
#> [29,] 0.7933097 0.9387413 0.7841721
#> [30,] 0.6263624 0.9343983 0.6452266
#> [31,] 0.6114372 0.9334051 0.6339973
#> [32,] 0.7497434 0.8954525 0.7573368
#> [33,] 0.7261246 0.9402274 0.7502049
#> [34,] 0.7551899 0.9557979 0.7701373
#> [35,] 0.8209651 0.9477411 0.8089738
#> [36,] 0.7411282 0.9402104 0.7446569
#> [37,] 0.5404581 0.8294197 0.5008452
#> [38,] 0.7970115 0.9689507 0.7944132
#> [39,] 0.7380407 0.8991771 0.7427898
#> [40,] 0.8201742 0.9486263 0.8161592
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6639673 0.3360327
#>  [2,] 0.6708104 0.3291896
#>  [3,] 0.6692487 0.3307513
#>  [4,] 0.6639249 0.3360751
#>  [5,] 0.6635345 0.3364655
#>  [6,] 0.6715119 0.3284881
#>  [7,] 0.6696677 0.3303323
#>  [8,] 0.6712255 0.3287745
#>  [9,] 0.6615976 0.3384024
#> [10,] 0.6677503 0.3322497
#> [11,] 0.6669907 0.3330093
#> [12,] 0.6677448 0.3322552
#> [13,] 0.6640321 0.3359679
#> [14,] 0.6689649 0.3310351
#> [15,] 0.6688580 0.3311420
#> [16,] 0.6659770 0.3340230
#> [17,] 0.6678795 0.3321205
#> [18,] 0.6550548 0.3449452
#> [19,] 0.6700189 0.3299811
#> [20,] 0.6656963 0.3343037
#> [21,] 0.6642499 0.3357501
#> [22,] 0.6707080 0.3292920
#> [23,] 0.6540199 0.3459801
#> [24,] 0.6663083 0.3336917
#> [25,] 0.6590277 0.3409723
#> [26,] 0.6638965 0.3361035
#> [27,] 0.6709490 0.3290510
#> [28,] 0.6666434 0.3333566
#> [29,] 0.6686424 0.3313576
#> [30,] 0.6606598 0.3393402
#> [31,] 0.6598292 0.3401708
#> [32,] 0.6637446 0.3362554
#> [33,] 0.6659594 0.3340406
#> [34,] 0.6677158 0.3322842
#> [35,] 0.6697861 0.3302139
#> [36,] 0.6666294 0.3333706
#> [37,] 0.6483125 0.3516875
#> [38,] 0.6701948 0.3298052
#> [39,] 0.6614234 0.3385766
#> [40,] 0.6707847 0.3292153
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]       [,3]         [,4]       [,5]         [,6]
#>  [1,]  0.17659029 1.4075184 -0.5209453  0.287457349 -0.4839984 -0.222930032
#>  [2,]  0.50788825 1.8499291 -1.0218647  0.330234025 -0.8307280 -0.267675442
#>  [3,]  0.42565899 1.7664134 -0.6135439  0.379782279 -0.5030922 -0.211673419
#>  [4,]  0.31979490 1.3549901 -1.6226176 -0.203209821 -1.3795820 -0.800189147
#>  [5,] -0.05543623 1.5565717 -0.9484716 -0.200334302 -0.8991992 -0.454890134
#>  [6,]  0.63353219 1.8891177 -1.1795802  0.224431462 -0.8508073 -0.218393969
#>  [7,]  0.18880102 1.9267504 -0.4306755  0.350804142 -0.3127245  0.005857279
#>  [8,]  0.66194854 1.8625061 -0.7720975  0.451555603 -0.5912771 -0.063799325
#>  [9,] -0.40170097 1.2202852 -1.0887763 -0.170406312 -0.9174983 -0.666216308
#> [10,] -0.01092742 1.7176888 -0.7379458  0.131703504 -0.6644399 -0.398693088
#> [11,]  0.44066010 1.5092302 -0.5001688  0.468089001 -0.3844533 -0.117171841
#> [12,]  0.63107469 1.6677502 -0.5863718  0.355101693 -0.4541886 -0.026256551
#> [13,]  0.19525826 1.2282517 -1.3531212 -0.038685954 -1.1005144 -0.570797495
#> [14,]  0.02306570 1.8765591 -0.7172339  0.140288200 -0.6454882 -0.385259215
#> [15,]  0.05215949 1.8577195 -0.6653451  0.229852836 -0.5973402 -0.321602908
#> [16,]  0.34524987 1.5331170 -0.8293227  0.267120681 -0.6679861 -0.278002233
#> [17,]  0.42163727 1.5949902 -0.6766912  0.364918285 -0.5575279 -0.254679754
#> [18,] -0.69252180 1.0120697 -1.0874033 -0.441674011 -1.1183559 -0.859533110
#> [19,]  0.61767394 1.7648072 -0.6538134  0.477132015 -0.4995791 -0.142629981
#> [20,]  0.28992274 1.6229662 -0.4755756  0.189769076 -0.4680196 -0.085577620
#> [21,] -0.08414511 1.3425838 -1.8713075 -0.326718190 -1.4412986 -0.857569524
#> [22,]  0.24552011 1.9384481 -0.7504247  0.306836954 -0.6696508 -0.337331988
#> [23,] -0.18452980 0.7630987 -1.6525042 -0.459857339 -1.4632021 -0.896263659
#> [24,] -0.10144700 1.5240452 -1.3519761  0.008541497 -1.1632227 -0.774037961
#> [25,] -0.13123183 0.9206090 -1.5283204 -0.307222979 -1.3507740 -0.969258952
#> [26,]  0.27770935 1.3262248 -1.7104022 -0.255490866 -1.4562098 -0.852420177
#> [27,]  0.18872205 2.1362430 -0.3053734  0.402220979 -0.2948993  0.018880068
#> [28,] -0.20222068 1.6971807 -0.9655621 -0.060009530 -0.8685051 -0.571710745
#> [29,]  0.47760494 1.6831137 -0.5358735  0.415956404 -0.4219874 -0.173187150
#> [30,] -0.61147324 1.3388084 -1.0539934 -0.353881203 -0.9572384 -0.758759122
#> [31,] -0.56394073 1.3591944 -0.9989051 -0.434028425 -1.0402933 -0.741921691
#> [32,]  0.53339193 1.3070121 -0.3284036  0.470364784 -0.2382545 -0.024029825
#> [33,]  0.04616336 1.5335578 -0.9629482  0.077186630 -0.8241282 -0.665867024
#> [34,] -0.04812124 1.8146802 -0.3739329  0.270609948 -0.3620741 -0.115817488
#> [35,]  0.47601813 1.8873765 -0.9181760  0.203381825 -0.7766367 -0.228384146
#> [36,]  0.21940995 1.6617161 -0.7028424  0.088293686 -0.5517697 -0.214598331
#> [37,] -0.39425110 0.4786366 -1.9566277 -1.021061105 -1.5211415 -1.190793501
#> [38,] -0.01186359 2.0854117 -0.4769740  0.250460193 -0.4660484 -0.097098728
#> [39,]  0.25598232 1.2284427 -1.6832210 -0.282810068 -1.4125378 -0.745188142
#> [40,]  0.60807493 1.9336995 -0.2715199  0.599547960 -0.1781082  0.082592944
#>              [,7]      [,8]     [,9]       [,10]
#>  [1,]  0.64190829 1.0502539 2.347167 1.073274184
#>  [2,]  0.66705702 1.5690334 2.965809 1.559747387
#>  [3,]  0.83351392 1.4090323 2.804749 1.417520021
#>  [4,]  0.11988123 1.1400560 2.415887 1.040779601
#>  [5,]  0.50905837 0.9935141 2.548094 0.968893539
#>  [6,]  0.56954802 1.6473082 3.061547 1.500373138
#>  [7,]  1.04210505 1.2506792 3.168316 1.314932202
#>  [8,]  0.74730809 1.6018556 2.951081 1.539155698
#>  [9,]  0.57236912 0.5549293 2.549189 0.711227596
#> [10,]  0.91217482 1.0892254 3.018541 1.163684948
#> [11,]  0.76953878 1.2445121 2.471826 1.262688822
#> [12,]  0.62833625 1.4354368 2.525775 1.350374144
#> [13,]  0.37541401 0.8180854 2.546607 0.949785782
#> [14,]  0.96663243 1.2045905 3.181604 1.202451833
#> [15,]  0.95342871 1.1994901 3.086430 1.200002010
#> [16,]  0.48235294 1.2499836 2.432278 1.209306976
#> [17,]  0.72504919 1.2770773 2.638621 1.208889056
#> [18,]  0.29504828 0.1254589 2.248487 0.305456829
#> [19,]  0.77002385 1.5122256 2.744487 1.399720591
#> [20,]  0.72197507 1.1826314 2.561371 1.074514675
#> [21,]  0.20221942 0.8049775 2.875766 0.801892338
#> [22,]  0.97391491 1.4062757 3.192249 1.393384236
#> [23,] -0.10006342 0.4429354 1.833357 0.559438899
#> [24,]  0.48141116 1.0411161 2.864569 1.098258711
#> [25,]  0.18867651 0.6141217 2.220525 0.696952327
#> [26,]  0.11888788 1.0986486 2.479768 1.008153321
#> [27,]  1.20696358 1.5227252 3.306284 1.473039540
#> [28,]  0.80879857 0.9525891 3.035078 0.974110753
#> [29,]  0.80169140 1.3449919 2.729433 1.290147316
#> [30,]  0.51468863 0.5166429 2.656302 0.598122359
#> [31,]  0.53293769 0.4533575 2.640211 0.549404137
#> [32,]  0.63805050 1.0972441 2.147688 1.138133710
#> [33,]  0.63887733 0.9750486 2.755575 1.099705583
#> [34,]  1.03496256 1.1264865 3.073775 1.209086536
#> [35,]  0.66953456 1.5229000 2.897872 1.443356045
#> [36,]  0.74124226 1.0518407 2.755272 1.070315313
#> [37,] -0.31667452 0.1621868 1.581520 0.003380786
#> [38,]  1.14412803 1.3677200 3.440637 1.351735700
#> [39,] -0.00303374 1.0358098 2.188115 1.060519277
#> [40,]  1.04618868 1.5175280 2.915888 1.490539148
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3619660 -0.3190586
#>  [2,] 0.3786988 -0.3331540
#>  [3,] 0.3773344 -0.3274547
#>  [4,] 0.3525519 -0.3282823
#>  [5,] 0.3573395 -0.3217457
#>  [6,] 0.3801623 -0.3348689
#>  [7,] 0.3794032 -0.3272794
#>  [8,] 0.3824687 -0.3312643
#>  [9,] 0.3460880 -0.3243338
#> [10,] 0.3704118 -0.3276159
#> [11,] 0.3718553 -0.3227503
#> [12,] 0.3742597 -0.3237432
#> [13,] 0.3520139 -0.3293011
#> [14,] 0.3750785 -0.3284285
#> [15,] 0.3750741 -0.3279506
#> [16,] 0.3652554 -0.3247901
#> [17,] 0.3727029 -0.3259072
#> [18,] 0.3235871 -0.3177461
#> [19,] 0.3801573 -0.3281133
#> [20,] 0.3692937 -0.3194900
#> [21,] 0.3480974 -0.3341938
#> [22,] 0.3807516 -0.3306375
#> [23,] 0.3218391 -0.3149173
#> [24,] 0.3589459 -0.3325891
#> [25,] 0.3363617 -0.3226024
#> [26,] 0.3515218 -0.3291855
#> [27,] 0.3869967 -0.3254835
#> [28,] 0.3648894 -0.3281529
#> [29,] 0.3759812 -0.3260700
#> [30,] 0.3428942 -0.3233415
#> [31,] 0.3412002 -0.3213329
#> [32,] 0.3638823 -0.3161440
#> [33,] 0.3605425 -0.3294235
#> [34,] 0.3735320 -0.3243401
#> [35,] 0.3774413 -0.3297766
#> [36,] 0.3677536 -0.3252260
#> [37,] 0.3071103 -0.3045194
#> [38,] 0.3823685 -0.3266978
#> [39,] 0.3447391 -0.3249047
#> [40,] 0.3857383 -0.3259977
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.6639673  0.6639673
#>  [2,] -0.6708104  0.6708104
#>  [3,] -0.6692487  0.6692487
#>  [4,]  0.3360751 -0.3360751
#>  [5,] -0.6635345  0.6635345
#>  [6,] -0.6715119  0.6715119
#>  [7,] -0.6696677  0.6696677
#>  [8,] -0.6712255  0.6712255
#>  [9,]  0.3384024 -0.3384024
#> [10,]  0.3322497 -0.3322497
#> [11,] -0.6669907  0.6669907
#> [12,] -0.6677448  0.6677448
#> [13,] -0.6640321  0.6640321
#> [14,]  0.3310351 -0.3310351
#> [15,]  0.3311420 -0.3311420
#> [16,] -0.6659770  0.6659770
#> [17,]  0.3321205 -0.3321205
#> [18,] -0.6550548  0.6550548
#> [19,]  0.3299811 -0.3299811
#> [20,] -0.6656963  0.6656963
#> [21,]  0.3357501 -0.3357501
#> [22,]  0.3292920 -0.3292920
#> [23,]  0.3459801 -0.3459801
#> [24,]  0.3336917 -0.3336917
#> [25,]  0.3409723 -0.3409723
#> [26,]  0.3361035 -0.3361035
#> [27,] -0.6709490  0.6709490
#> [28,]  0.3333566 -0.3333566
#> [29,]  0.3313576 -0.3313576
#> [30,]  0.3393402 -0.3393402
#> [31,]  0.3401708 -0.3401708
#> [32,] -0.6637446  0.6637446
#> [33,]  0.3340406 -0.3340406
#> [34,] -0.6677158  0.6677158
#> [35,] -0.6697861  0.6697861
#> [36,] -0.6666294  0.6666294
#> [37,]  0.3516875 -0.3516875
#> [38,] -0.6701948  0.6701948
#> [39,] -0.6614234  0.6614234
#> [40,] -0.6707847  0.6707847
#> 
#> $L
#> [1] 0.6978353 0.6765007 0.6908415 0.6912401 0.7169297 0.7717802
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609