Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10           V11          V12         V13         V14
#>  [1,] -0.3059918 -0.07270277 -0.0540356810 -0.103541364 -0.11082848 -0.04276335
#>  [2,] -0.2783087 -0.12150941 -0.1903476921 -0.181964870 -0.11231101 -0.00693931
#>  [3,] -0.3775112 -0.17807214 -0.1016630237 -0.046373090 -0.04776449 -0.04509942
#>  [4,] -0.2183289 -0.07175758 -0.0009629799 -0.064079952 -0.07820924 -0.03585712
#>  [5,] -0.2429945 -0.19373258 -0.0756222405 -0.118717745 -0.17614346 -0.05494899
#>  [6,] -0.1564871 -0.03603267 -0.0638361753 -0.163750677 -0.12301691 -0.09886028
#>  [7,] -0.3039379 -0.07426473 -0.1527485029 -0.015372244  0.06286158  0.06509905
#>  [8,] -0.4415651 -0.13692759  0.0201165992  0.002004296  0.08087088 -0.04563156
#>  [9,] -0.3401912 -0.07631466 -0.0737075077 -0.072333091 -0.03810092 -0.11742405
#> [10,] -0.2183287 -0.11107039 -0.1300186004 -0.129052512 -0.13104331 -0.15591136
#>                V15          V16         V17         V18         V19         V2
#>  [1,] -0.100056346  0.012504767 -0.06301146 -0.01217210  0.11240694 -0.2687438
#>  [2,] -0.069031228 -0.094239026  0.02085560  0.00746694  0.03958159 -0.2813574
#>  [3,]  0.032094242  0.015557381 -0.02509734 -0.01475636  0.12221417 -0.2495624
#>  [4,]  0.052150493 -0.060529498 -0.07223433  0.01486179  0.10690783 -0.3914356
#>  [5,] -0.104832487 -0.183486021 -0.03042087 -0.02823053 -0.12817993 -0.3155771
#>  [6,] -0.025491870 -0.048834064 -0.01346115 -0.09005284 -0.05099041 -0.3244956
#>  [7,]  0.007045477  0.041275983  0.07005747 -0.05160460 -0.05110083 -0.2528759
#>  [8,] -0.037767645  0.020887633  0.21334893  0.06713986  0.20397848 -0.3610945
#>  [9,] -0.138268401  0.050214112 -0.07278113 -0.01944476 -0.07831587 -0.2314471
#> [10,]  0.007890913 -0.004357873  0.01757329  0.15975565  0.03856283 -0.3161608
#>                V20         V21        V22        V23         V24         V25
#>  [1,]  0.179379953  0.16580420 0.13020463 0.04735897  0.09842464  0.05549287
#>  [2,] -0.007680689 -0.10358176 0.06609316 0.10155125 -0.01762596 -0.06183118
#>  [3,]  0.116400948  0.12620005 0.12829236 0.18492802  0.12837910  0.02444228
#>  [4,]  0.072337688  0.21716547 0.07967872 0.02222131  0.05901848  0.08713296
#>  [5,] -0.124894548 -0.00715763 0.08892266 0.06832194  0.03090898  0.13726860
#>  [6,]  0.123935043  0.01227634 0.01781288 0.07268611 -0.03062857  0.01385760
#>  [7,]  0.047214170  0.17167600 0.11427434 0.12236374  0.07586885  0.11713361
#>  [8,]  0.171429900  0.22853517 0.20682468 0.19443677  0.34581603  0.29510829
#>  [9,]  0.098733619  0.06944003 0.13189325 0.09386630  0.21717504  0.23565169
#> [10,]  0.023651985  0.12371498 0.19475855 0.19298085  0.07402433  0.16340541
#>               V26         V27         V28         V29         V3          V30
#>  [1,]  0.20284818  0.20217703 0.127496216  0.14005513 -0.2338563  0.090385314
#>  [2,]  0.05731000  0.07094306 0.044937378  0.04452545 -0.2752877 -0.017949205
#>  [3,]  0.20263652  0.19184785 0.208858365  0.16657553 -0.3634037  0.037062891
#>  [4,]  0.12213163  0.18025461 0.116345431  0.20939749 -0.2453206  0.117136840
#>  [5,]  0.17314378  0.12637227 0.098666609  0.13117119 -0.3046599  0.083023778
#>  [6,]  0.03138845 -0.06713835 0.075169339 -0.06709197 -0.3097149  0.006350218
#>  [7,]  0.13108021  0.20083408 0.182023542  0.04751288 -0.2128371  0.118891004
#>  [8,]  0.32357140  0.24706064 0.196473166  0.33538261 -0.2694703  0.245299978
#>  [9,]  0.21952613  0.11102460 0.198535002  0.09414153 -0.3195783  0.134683402
#> [10,] -0.02856710  0.13780741 0.007452876  0.02259452 -0.3059889  0.034647072
#>                V31          V32         V33         V34         V35
#>  [1,]  0.037275088 -0.014005737 -0.01387705  0.01800252 -0.01346507
#>  [2,]  0.020192406 -0.051795051  0.07930270  0.05206278 -0.01482711
#>  [3,]  0.097532995 -0.021640148 -0.02911718 -0.07236946  0.01027725
#>  [4,]  0.096768497 -0.003569216  0.07633491 -0.02686376  0.06801187
#>  [5,]  0.075948601  0.011128747 -0.05256037 -0.14733099 -0.05436300
#>  [6,] -0.118568983  0.002401975 -0.10510439  0.04739029 -0.13110266
#>  [7,]  0.018931174 -0.082733934  0.01287362 -0.03259971 -0.03871160
#>  [8,]  0.237713656  0.162735536  0.11752729  0.07344555 -0.01363620
#>  [9,] -0.009091575  0.082231901  0.05678322 -0.06462995  0.05252591
#> [10,]  0.007362097 -0.089626417 -0.05362863 -0.08581566 -0.10229177
#>                 V36         V37         V38         V39         V4         V40
#>  [1,] -0.1342700385 -0.02798965 -0.08940122 -0.10752899 -0.1878787 -0.14274020
#>  [2,]  0.1109457550 -0.13363598 -0.11215642  0.01485881 -0.3655682 -0.10883409
#>  [3,]  0.0188434859  0.01642517 -0.08100572 -0.08230842 -0.2580890  0.01542011
#>  [4,] -0.0001814448  0.06571110  0.09712827  0.08404919 -0.2799699 -0.09388679
#>  [5,] -0.1056865073 -0.20304268 -0.07614938 -0.09333827 -0.2445704 -0.15628487
#>  [6,] -0.0319327936 -0.10988976 -0.14594418 -0.06455606 -0.1861753 -0.15287566
#>  [7,] -0.0066068044 -0.08938822 -0.05471615 -0.11583161 -0.2585445  0.01406574
#>  [8,]  0.0151248280 -0.04330451 -0.03402851 -0.09147407 -0.2953656  0.06307795
#>  [9,] -0.0867524000 -0.03281972  0.01788389  0.02151899 -0.3272491 -0.10185186
#> [10,] -0.1191865993 -0.05005062 -0.17674507 -0.16674817 -0.2082480 -0.08592604
#>               V41         V42         V43          V44         V45         V46
#>  [1,]  0.02085999 -0.06930341 -0.05187731 -0.155252323 -0.16891706 -0.02005792
#>  [2,] -0.10776858 -0.04249869 -0.16196930 -0.076313428 -0.21970436 -0.15251871
#>  [3,] -0.03604550 -0.01457460 -0.02117342 -0.032901433 -0.13921108 -0.20797503
#>  [4,] -0.06846057 -0.09534820  0.01114874 -0.116771898 -0.18698503 -0.11791743
#>  [5,] -0.08408723 -0.04410179 -0.02400692 -0.159883235 -0.19429193 -0.16575830
#>  [6,] -0.14985669 -0.17564682 -0.17518863 -0.079567716 -0.11531215 -0.20224784
#>  [7,] -0.17510258 -0.06497816 -0.04157684 -0.093675142 -0.06241559 -0.05558245
#>  [8,]  0.03966384 -0.01458286 -0.11002477 -0.005394814 -0.01959637 -0.15692287
#>  [9,] -0.10397408 -0.13746283 -0.13163439 -0.192264807 -0.04705145 -0.09382503
#> [10,] -0.18630559 -0.20850656 -0.20484121 -0.184502647 -0.12355258 -0.24519223
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.2008698 -0.1101911 -0.1611612 -0.1905835 -0.2237932 -0.3702622
#>  [2,] -0.1622575 -0.1919565 -0.2257073 -0.1863602 -0.3723897 -0.2375162
#>  [3,] -0.2334939 -0.3506040 -0.2986321 -0.3250537 -0.4293067 -0.4475782
#>  [4,] -0.2261775 -0.2217606 -0.2865817 -0.1647289 -0.4119912 -0.3735790
#>  [5,] -0.1290069 -0.2966512 -0.3077930 -0.2815825 -0.2276219 -0.3668256
#>  [6,] -0.2468398 -0.2887886 -0.2023633 -0.1545008 -0.2416169 -0.1940566
#>  [7,] -0.1128232 -0.2673077 -0.2638352 -0.1839050 -0.2918060 -0.3513217
#>  [8,] -0.1213339 -0.1621417 -0.3403078 -0.2908227 -0.3184679 -0.3478377
#>  [9,] -0.2023519 -0.1638193 -0.1550304 -0.1479099 -0.3394033 -0.2922722
#> [10,] -0.1345047 -0.1634245 -0.2691673 -0.2052808 -0.4015300 -0.3319007
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.2799440 -0.3065478 -0.3244135 -0.3864494 -0.3151820 -0.3347483
#>  [2,] -0.3999678 -0.4105335 -0.3037371 -0.3358518 -0.3139206 -0.3025328
#>  [3,] -0.3869787 -0.3399238 -0.3268716 -0.3786917 -0.4208658 -0.4595401
#>  [4,] -0.3521833 -0.3051800 -0.4657128 -0.3969184 -0.3632301 -0.3620181
#>  [5,] -0.3233206 -0.4099420 -0.2363412 -0.2141496 -0.2361718 -0.2745065
#>  [6,] -0.3032902 -0.2216401 -0.2504260 -0.2475867 -0.3043099 -0.3509394
#>  [7,] -0.3357976 -0.4240438 -0.4402672 -0.3649559 -0.2809639 -0.4057851
#>  [8,] -0.3320233 -0.4636737 -0.3258499 -0.4134216 -0.4517687 -0.3874844
#>  [9,] -0.3515968 -0.2150845 -0.3206703 -0.2266107 -0.2898832 -0.2604388
#> [10,] -0.3182513 -0.2465725 -0.3755682 -0.3956551 -0.2914969 -0.2677778
#>              V58        V59          V6        V60          V7          V8
#>  [1,] -0.3698750 -0.3995721 -0.08090565 -0.2576481 -0.25649969 -0.26948471
#>  [2,] -0.3294091 -0.2938851 -0.22780551 -0.2879136 -0.14094010 -0.19967040
#>  [3,] -0.3314080 -0.3428498 -0.19235566 -0.4721037 -0.21625680 -0.31125959
#>  [4,] -0.3013064 -0.2762696 -0.21867755 -0.3318331 -0.20463120 -0.08845087
#>  [5,] -0.3658864 -0.2372298 -0.16435480 -0.2298353 -0.17653229 -0.15718937
#>  [6,] -0.2760702 -0.3210816 -0.09756630 -0.3427108 -0.12223387 -0.11558934
#>  [7,] -0.2510509 -0.4255836 -0.29346291 -0.2797354 -0.12328341 -0.19637359
#>  [8,] -0.4028831 -0.3824400 -0.22341057 -0.5057452 -0.16254073 -0.11301314
#>  [9,] -0.2637514 -0.3888633 -0.11736260 -0.2747140 -0.06653374 -0.05971567
#> [10,] -0.3174757 -0.2492085 -0.21122272 -0.2931244 -0.11380705 -0.15929364
#>                 V9
#>  [1,] -0.005443862
#>  [2,] -0.224290650
#>  [3,] -0.122289664
#>  [4,] -0.035840434
#>  [5,] -0.209603195
#>  [6,] -0.166680508
#>  [7,] -0.188704169
#>  [8,] -0.122429503
#>  [9,] -0.014947993
#> [10,] -0.168598076
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.1058451 -0.06286840 -0.01808545  0.07621990 -0.08423723 -0.08466661
#> [2,] -0.1274070  0.04957896 -0.15285949 -0.05479802  0.01195275  0.01999971
#>             [,7]       [,8]        [,9]       [,10]
#> [1,]  0.10165068  0.2049335 0.099384334 -0.05352279
#> [2,] -0.09335433 -0.1118545 0.003089368  0.05875256
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -2.615398e-04 -1.296592e-03 -1.457525e-03 -1.393305e-03 -1.189609e-03
#>  [2,]  2.213447e-04  1.289415e-03  1.449381e-03  1.408843e-03  1.286719e-03
#>  [3,] -7.472221e-05 -3.035796e-04 -3.359897e-04 -3.239435e-04 -2.332909e-04
#>  [4,] -5.196268e-05 -1.859350e-04 -2.132327e-04 -2.228617e-04 -1.433421e-04
#>  [5,]  1.803432e-04  1.074587e-03  1.225471e-03  1.213071e-03  1.112440e-03
#>  [6,]  2.008812e-04  1.189738e-03  1.352738e-03  1.306150e-03  1.176327e-03
#>  [7,] -2.026187e-04 -9.958234e-04 -1.113458e-03 -1.061859e-03 -8.936979e-04
#>  [8,] -1.112069e-04 -4.587041e-04 -4.804130e-04 -3.917756e-04 -3.047000e-04
#>  [9,] -2.234202e-05 -3.434441e-05 -4.067688e-05 -5.403469e-05  1.583475e-05
#> [10,]  2.775604e-04  1.602802e-03  1.845974e-03  1.800646e-03  1.672582e-03
#>                 V14           V15           V16           V17           V18
#>  [1,] -1.307090e-03 -9.923127e-04 -8.611846e-04 -5.246878e-04 -7.367922e-04
#>  [2,]  1.419525e-03  1.247839e-03  1.219617e-03  9.977639e-04  1.271107e-03
#>  [3,] -2.632643e-04 -1.259540e-04 -5.972502e-05  7.349719e-05  4.239628e-05
#>  [4,] -1.501414e-04 -2.821834e-05  2.847902e-05  1.301240e-04  1.115203e-04
#>  [5,]  1.195753e-03  1.034820e-03  1.012678e-03  8.319327e-04  1.066813e-03
#>  [6,]  1.253463e-03  1.096220e-03  1.030181e-03  8.124820e-04  1.081274e-03
#>  [7,] -9.842034e-04 -7.315598e-04 -6.244917e-04 -3.531404e-04 -5.035632e-04
#>  [8,] -4.620616e-04 -3.952798e-04 -3.473594e-04 -1.792866e-04 -2.257887e-04
#>  [9,]  2.835257e-05  1.280852e-04  1.844680e-04  2.648861e-04  2.737897e-04
#> [10,]  1.773731e-03  1.561812e-03  1.505976e-03  1.263545e-03  1.611770e-03
#>                 V19            V2           V20           V21           V22
#>  [1,] -0.0021102587 -2.660944e-04 -2.762996e-03 -2.494859e-03 -2.456129e-03
#>  [2,]  0.0024174627  2.417341e-04  2.900960e-03  2.688725e-03  2.769538e-03
#>  [3,] -0.0003554965 -6.512620e-05 -5.749563e-04 -4.727048e-04 -4.457147e-04
#>  [4,] -0.0001798652 -4.122303e-05 -3.344773e-04 -2.444914e-04 -2.450146e-04
#>  [5,]  0.0020490910  2.006327e-04  2.421889e-03  2.241757e-03  2.374348e-03
#>  [6,]  0.0021969925  2.236317e-04  2.637008e-03  2.431973e-03  2.477089e-03
#>  [7,] -0.0015594224 -2.047284e-04 -2.078969e-03 -1.868350e-03 -1.840016e-03
#>  [8,] -0.0007014609 -1.058236e-04 -1.014010e-03 -9.100188e-04 -7.564401e-04
#>  [9,]  0.0001177458 -8.468355e-06  2.211764e-05  8.638374e-05  9.024556e-05
#> [10,]  0.0031014178  3.168369e-04  3.671944e-03  3.455310e-03  3.521313e-03
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0022628065 -0.0017174873 -1.071397e-03 -1.038250e-03 -8.187110e-04
#>  [2,]  0.0027045359  0.0022744804  1.732363e-03  1.715173e-03  1.519547e-03
#>  [3,] -0.0003779943 -0.0001926885 -1.667955e-06 -6.088849e-06  3.929465e-05
#>  [4,] -0.0001998524 -0.0000446058  1.030939e-04  1.042879e-04  1.259286e-04
#>  [5,]  0.0023621924  0.0019906947  1.526655e-03  1.505560e-03  1.357506e-03
#>  [6,]  0.0023870681  0.0019506765  1.391435e-03  1.340951e-03  1.155471e-03
#>  [7,] -0.0016815949 -0.0012431402 -7.397649e-04 -7.119195e-04 -5.503527e-04
#>  [8,] -0.0006115881 -0.0004744068 -2.619956e-04 -2.635123e-04 -1.765801e-04
#>  [9,]  0.0001300062  0.0002465772  3.341153e-04  3.318959e-04  3.392433e-04
#> [10,]  0.0034191551  0.0028984925  2.179487e-03  2.098452e-03  1.836897e-03
#>                 V28           V29            V3           V30           V31
#>  [1,] -4.068878e-04 -4.773061e-04 -3.012262e-04 -8.576395e-04 -3.166193e-04
#>  [2,]  1.168598e-03  1.161324e-03  2.677179e-04  1.396832e-03  9.103241e-04
#>  [3,]  1.710337e-04  1.257911e-04 -7.537131e-05 -3.164396e-05  1.024585e-04
#>  [4,]  2.278697e-04  1.873400e-04 -4.949824e-05  6.745818e-05  1.450515e-04
#>  [5,]  1.056058e-03  1.035278e-03  2.247895e-04  1.197773e-03  7.964344e-04
#>  [6,]  8.608908e-04  8.688663e-04  2.465567e-04  1.062511e-03  5.969255e-04
#>  [7,] -2.376825e-04 -3.030771e-04 -2.308312e-04 -6.033751e-04 -1.787071e-04
#>  [8,]  9.586805e-06 -7.122084e-05 -1.225986e-04 -3.129092e-04 -9.996399e-05
#>  [9,]  4.013903e-04  3.533030e-04 -1.150187e-05  2.586716e-04  2.854676e-04
#> [10,]  1.430124e-03  1.384346e-03  3.528048e-04  1.593867e-03  9.582957e-04
#>                 V32           V33           V34           V35           V36
#>  [1,] -5.870063e-04 -6.166979e-04 -2.770200e-04  3.314825e-04  3.348929e-04
#>  [2,]  9.709033e-04  9.032497e-04  5.147513e-04  1.572855e-05  9.491043e-05
#>  [3,] -2.290956e-05 -5.648722e-05  2.028208e-05  2.111860e-04  2.358369e-04
#>  [4,]  3.445318e-05  7.064579e-06  6.188827e-05  2.103908e-04  2.224361e-04
#>  [5,]  8.509446e-04  7.660054e-04  3.968822e-04 -5.308846e-05  4.083027e-05
#>  [6,]  7.188070e-04  6.710609e-04  3.127132e-04 -1.557139e-04 -6.898594e-05
#>  [7,] -4.024358e-04 -4.203700e-04 -1.619088e-04  3.091678e-04  3.290381e-04
#>  [8,] -1.549051e-04 -2.312704e-04 -1.946754e-04 -3.690018e-05  4.296898e-05
#>  [9,]  1.812333e-04  1.435927e-04  1.576946e-04  2.510118e-04  2.646587e-04
#> [10,]  1.110808e-03  1.035558e-03  5.621000e-04 -7.861979e-05  4.773393e-06
#>                 V37           V38           V39            V4           V40
#>  [1,] -3.502329e-04 -8.168341e-04 -1.007657e-03 -4.049177e-04 -8.429498e-04
#>  [2,]  6.586181e-04  1.003032e-03  1.174996e-03  3.541533e-04  1.026398e-03
#>  [3,]  1.774076e-05 -1.206655e-04 -1.542917e-04 -1.001476e-04 -1.219752e-04
#>  [4,]  4.490810e-05 -4.986222e-05 -6.293097e-05 -6.112449e-05 -4.687927e-05
#>  [5,]  5.350415e-04  8.229632e-04  9.904188e-04  2.961417e-04  8.785104e-04
#>  [6,]  5.009000e-04  8.560172e-04  1.055453e-03  3.250924e-04  9.192803e-04
#>  [7,] -2.104741e-04 -5.850274e-04 -7.324090e-04 -3.110640e-04 -6.132764e-04
#>  [8,] -1.043108e-04 -2.555555e-04 -2.558386e-04 -1.681969e-04 -1.819211e-04
#>  [9,]  1.378025e-04  7.955915e-05  8.200596e-05 -1.208067e-05  7.189925e-05
#> [10,]  7.158813e-04  1.205462e-03  1.507536e-03  4.735132e-04  1.313569e-03
#>                 V41           V42           V43           V44           V45
#>  [1,] -1.140940e-03 -1.378031e-03 -1.378511e-03 -1.426182e-03 -0.0017317958
#>  [2,]  1.211258e-03  1.362069e-03  1.355369e-03  1.338422e-03  0.0014959321
#>  [3,] -2.234440e-04 -3.084615e-04 -3.193256e-04 -3.455424e-04 -0.0004537656
#>  [4,] -1.217218e-04 -1.960723e-04 -2.093608e-04 -2.252323e-04 -0.0002974898
#>  [5,]  1.050681e-03  1.198064e-03  1.188951e-03  1.160298e-03  0.0012864771
#>  [6,]  1.096934e-03  1.263250e-03  1.244954e-03  1.222029e-03  0.0013917149
#>  [7,] -8.634672e-04 -1.058409e-03 -1.052614e-03 -1.090510e-03 -0.0013315782
#>  [8,] -2.603222e-04 -3.021262e-04 -3.214144e-04 -4.061272e-04 -0.0005427009
#>  [9,]  2.099464e-05 -3.236359e-05 -4.931948e-05 -6.381303e-05 -0.0001129381
#> [10,]  1.587546e-03  1.801665e-03  1.756596e-03  1.724535e-03  0.0019645897
#>                 V46           V47           V48           V49            V5
#>  [1,] -1.498864e-03 -9.055115e-04 -6.158157e-04 -3.981502e-04 -4.721843e-04
#>  [2,]  1.312362e-03  8.497350e-04  5.551659e-04  3.679335e-04  4.496412e-04
#>  [3,] -3.893891e-04 -2.078440e-04 -1.476643e-04 -9.189306e-05 -1.108098e-04
#>  [4,] -2.564177e-04 -1.293878e-04 -9.043741e-05 -5.581351e-05 -7.056476e-05
#>  [5,]  1.135205e-03  7.500543e-04  4.853382e-04  3.224522e-04  3.845306e-04
#>  [6,]  1.228554e-03  8.100614e-04  5.256902e-04  3.542931e-04  4.089669e-04
#>  [7,] -1.154252e-03 -6.936715e-04 -4.733479e-04 -3.053961e-04 -3.623349e-04
#>  [8,] -4.538563e-04 -2.397527e-04 -1.787411e-04 -1.025842e-04 -1.660840e-04
#>  [9,] -9.531558e-05 -2.374036e-05 -1.907560e-05 -1.182588e-05 -1.384138e-05
#> [10,]  1.727045e-03  1.151506e-03  7.564741e-04  5.044562e-04  5.828448e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -1.038217e-04 -1.065296e-04 -1.147073e-04 -7.420183e-05 -7.680487e-05
#>  [2,]  1.024336e-04  9.874308e-05  1.048096e-04  6.899573e-05  6.658391e-05
#>  [3,] -2.131893e-05 -2.415264e-05 -2.940872e-05 -1.990156e-05 -2.114176e-05
#>  [4,] -1.110857e-05 -1.434720e-05 -1.956952e-05 -1.374183e-05 -1.477556e-05
#>  [5,]  8.820001e-05  8.606217e-05  8.945051e-05  5.812276e-05  5.565348e-05
#>  [6,]  9.470033e-05  9.407030e-05  9.601322e-05  6.081296e-05  6.041677e-05
#>  [7,] -7.861100e-05 -8.200358e-05 -8.869422e-05 -5.711220e-05 -5.921956e-05
#>  [8,] -2.762412e-05 -3.101794e-05 -4.052391e-05 -2.530387e-05 -2.614466e-05
#>  [9,]  1.594596e-06 -1.592944e-06 -6.354342e-06 -5.097717e-06 -6.241626e-06
#> [10,]  1.389610e-04  1.357206e-04  1.340230e-04  8.510406e-05  8.496882e-05
#>                 V55           V56           V57           V58           V59
#>  [1,] -2.984812e-05 -3.714004e-05 -3.332181e-05 -6.786579e-05 -6.010303e-05
#>  [2,]  2.977330e-05  3.261811e-05  3.051015e-05  6.115999e-05  5.445441e-05
#>  [3,] -5.521023e-06 -1.022733e-05 -8.433899e-06 -1.769627e-05 -1.580964e-05
#>  [4,] -2.409625e-06 -6.902745e-06 -5.420348e-06 -1.239030e-05 -1.105479e-05
#>  [5,]  2.396890e-05  2.669865e-05  2.473925e-05  5.189169e-05  4.619628e-05
#>  [6,]  2.549525e-05  2.839100e-05  2.529848e-05  5.720410e-05  5.007984e-05
#>  [7,] -2.194675e-05 -2.869944e-05 -2.526833e-05 -5.229321e-05 -4.636991e-05
#>  [8,] -1.006487e-05 -1.319862e-05 -1.513396e-05 -2.131350e-05 -2.092923e-05
#>  [9,]  2.317545e-06 -2.167658e-06 -2.683167e-07 -4.435192e-06 -3.832747e-06
#> [10,]  4.040028e-05  4.147677e-05  3.722444e-05  7.939705e-05  7.016615e-05
#>                  V6           V60            V7            V8            V9
#>  [1,] -3.917731e-04 -2.866730e-05 -3.278926e-04 -3.746357e-04 -1.003200e-03
#>  [2,]  4.706711e-04  2.692349e-05  4.362842e-04  4.460717e-04  9.922550e-04
#>  [3,] -6.026184e-05 -7.226116e-06 -2.568717e-05 -5.197217e-05 -2.366183e-04
#>  [4,] -2.938449e-05 -4.624359e-06  6.600904e-06 -1.369845e-05 -1.430400e-04
#>  [5,]  4.079817e-04  2.206516e-05  3.723952e-04  3.754534e-04  8.228838e-04
#>  [6,]  4.213986e-04  2.358883e-05  3.927001e-04  4.060183e-04  8.936897e-04
#>  [7,] -2.918976e-04 -2.202549e-05 -2.380456e-04 -2.800436e-04 -7.684509e-04
#>  [8,] -1.236815e-04 -1.222950e-05 -1.034721e-04 -1.159732e-04 -3.953254e-04
#>  [9,]  2.941476e-05 -9.236414e-07  6.279657e-05  4.224936e-05 -1.855496e-05
#> [10,]  5.941201e-04  3.363479e-05  5.725799e-04  5.703531e-04  1.207707e-03
#> 
#> $vW[[2]]
#>             [,1]         [,2]        [,3]       [,4]         [,5]        [,6]
#> [1,] -0.03387101 -0.008963361 -0.03641796 -0.0394119 -0.008734206 -0.01012459
#> [2,]  0.03387101  0.008963361  0.03641796  0.0394119  0.008734206  0.01012459
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.03229734 -0.05535502 -0.03366147 -0.01952693
#> [2,]  0.03229734  0.05535502  0.03366147  0.01952693
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.3600985 0.3284609 0.5645027 0.6612101 0.2407549 0.3312949 0.3928138
#>  [8] 0.8389142 0.3926076 0.3097743
#> 
#> $B[[2]]
#> [1]  0.07023784 -0.16490866
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -1.619953e-03  2.445734e-03 -6.444378e-05  1.060447e-04  2.080215e-03
#>  [6]  1.989127e-03 -1.136567e-03 -5.106254e-04  4.406825e-04  3.047603e-03
#> 
#> $vB[[2]]
#> [1] -0.06001148  0.06001148
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#>  [2,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#>  [3,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#>  [4,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#>  [5,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#>  [6,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [7,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [8,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#>  [9,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [10,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [11,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [12,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [13,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [14,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [15,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [16,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [17,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [18,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [19,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [20,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [21,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [22,] 0.0099 0.1882 0.1456 0.1892 0.3176 0.1340 0.2169 0.2458 0.2589 0.2786
#> [23,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [24,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [25,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [26,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [27,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [28,] 0.0654 0.3422 0.2128 0.1377 0.4032 0.5684 0.2398 0.4331 0.5954 0.5772
#> [29,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [30,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [31,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [32,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [33,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [34,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [35,] 0.1150 0.5378 0.3816 0.0991 0.0616 0.1795 0.3907 0.3602 0.3041 0.2428
#> [36,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [37,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [38,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#> [39,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [40,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#>  [2,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#>  [3,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#>  [4,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#>  [5,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#>  [6,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [7,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [8,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#>  [9,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [10,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [11,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [12,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [13,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [14,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [15,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [16,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [17,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [18,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [19,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [20,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [21,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [22,] 0.2298 0.0484 0.0656 0.1441 0.1179 0.1668 0.1783 0.2476 0.2570 0.1036
#> [23,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [24,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [25,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [26,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [27,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [28,] 0.8176 0.0649 0.8835 0.5248 0.6373 0.8375 0.6699 0.7756 0.8750 0.8300
#> [29,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [30,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [31,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [32,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [33,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [34,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [35,] 0.4060 0.1163 0.8395 0.9777 0.4680 0.0610 0.2143 0.1348 0.2854 0.1617
#> [36,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [37,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [38,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#> [39,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [40,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#>  [2,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#>  [3,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#>  [4,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#>  [5,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#>  [6,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [7,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [8,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#>  [9,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [10,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [11,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [12,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [13,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [14,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [15,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [16,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [17,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [18,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [19,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [20,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [21,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [22,] 0.5356 0.7124 0.0299 0.6291 0.4756 0.6015 0.7208 0.6234 0.5725 0.7523
#> [23,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [24,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [25,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [26,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [27,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [28,] 0.6896 0.3372 0.0737 0.6405 0.7138 0.8202 0.6657 0.5254 0.2960 0.0704
#> [29,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [30,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [31,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [32,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [33,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [34,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [35,] 0.2649 0.4565 0.0866 0.6502 0.2848 0.3296 0.5370 0.6627 0.8626 0.8547
#> [36,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [37,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [38,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#> [39,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [40,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#>  [2,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#>  [3,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#>  [4,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#>  [5,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#>  [6,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [7,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [8,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#>  [9,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [10,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [11,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [12,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [13,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [14,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [15,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [16,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [17,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [18,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [19,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [20,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [21,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [22,] 0.8712 0.9252 0.9709 0.0297 0.9297 0.8995 0.7911 0.5600 0.2838 0.4407
#> [23,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [24,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [25,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [26,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [27,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [28,] 0.0970 0.3941 0.6028 0.1132 0.3521 0.3924 0.4808 0.4602 0.4164 0.5438
#> [29,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [30,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [31,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [32,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [33,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [34,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [35,] 0.7848 0.9016 0.8827 0.0358 0.6086 0.2810 0.0906 0.1177 0.2694 0.5214
#> [36,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [37,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [38,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#> [39,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [40,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#>  [2,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#>  [3,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#>  [4,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#>  [5,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#>  [6,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [7,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [8,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#>  [9,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [10,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [11,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [12,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [13,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [14,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [15,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [16,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [17,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [18,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [19,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [20,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [21,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [22,] 0.5507 0.4331 0.2905 0.1981 0.0652 0.0779 0.0396 0.0173 0.0149 0.0115
#> [23,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [24,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [25,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [26,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [27,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [28,] 0.5649 0.3195 0.2484 0.1299 0.2482 0.0825 0.0243 0.0210 0.0361 0.0239
#> [29,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [30,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [31,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [32,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [33,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [34,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [35,] 0.4232 0.2340 0.1928 0.1092 0.0232 0.0507 0.0228 0.0099 0.0065 0.0085
#> [36,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [37,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [38,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#> [39,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [40,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#>  [2,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#>  [3,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#>  [4,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#>  [5,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#>  [6,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [7,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [8,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#>  [9,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [10,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [11,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [12,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [13,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [14,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [15,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [16,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [17,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [18,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [19,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [20,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [21,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [22,] 0.0202 0.0139 0.0029 0.0160 0.0106 0.1077 0.0134 0.2363 0.2385 0.0075
#> [23,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [24,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [25,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [26,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [27,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [28,] 0.0447 0.0394 0.0355 0.0440 0.0243 0.1257 0.0098 0.1797 0.0989 0.2460
#> [29,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [30,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [31,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [32,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [33,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [34,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [35,] 0.0166 0.0110 0.0190 0.0141 0.0068 0.1267 0.0086 0.2417 0.2661 0.4346
#> [36,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [37,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [38,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#> [39,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [40,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7460144 0.5271419 0.8138878 0.8288425 0.5940686 0.4444389 0.7500016
#>  [2,] 0.6896436 0.4317435 0.7748372 0.7858927 0.4562404 0.3848321 0.7206749
#>  [3,] 0.6511538 0.4445245 0.7441510 0.7657220 0.3603478 0.3812987 0.6621341
#>  [4,] 0.6707660 0.4531820 0.7299978 0.7650770 0.4909552 0.3666691 0.6632717
#>  [5,] 0.6255977 0.4634850 0.7136879 0.7769945 0.3864509 0.3610860 0.6301278
#>  [6,] 0.7539157 0.5115245 0.8155874 0.8228205 0.6046400 0.4528049 0.7457919
#>  [7,] 0.6991064 0.5100488 0.7611819 0.7755351 0.4342562 0.4540981 0.6777804
#>  [8,] 0.6915821 0.4539946 0.7489030 0.7665484 0.4003232 0.3802477 0.6829255
#>  [9,] 0.5896720 0.4022254 0.6771277 0.7040623 0.2910765 0.3569970 0.5838434
#> [10,] 0.7273570 0.5582718 0.8087450 0.8214242 0.5995177 0.4372135 0.7346727
#> [11,] 0.6066981 0.3851057 0.7002834 0.7236949 0.3298092 0.3391826 0.6466639
#> [12,] 0.5778718 0.3995927 0.6803134 0.7688499 0.3708764 0.2689348 0.5802983
#> [13,] 0.6660804 0.4089051 0.7505606 0.7657655 0.4323131 0.3602721 0.6974872
#> [14,] 0.6980916 0.5010553 0.7700273 0.7641588 0.4662960 0.4438763 0.6885099
#> [15,] 0.6123558 0.4377855 0.7170527 0.7531334 0.4188334 0.3792663 0.6520239
#> [16,] 0.7539521 0.5171803 0.8171044 0.8040016 0.5193964 0.4776713 0.7620400
#> [17,] 0.6886757 0.4594707 0.7686933 0.7760734 0.3993482 0.3918775 0.6806034
#> [18,] 0.6763815 0.5240996 0.7695525 0.7919369 0.5101246 0.4155162 0.6811244
#> [19,] 0.6815077 0.4801184 0.7629339 0.7916701 0.4261172 0.4002971 0.6592803
#> [20,] 0.5223668 0.3006619 0.5710118 0.7273922 0.2166518 0.2543630 0.4951219
#> [21,] 0.6676783 0.5305095 0.7618731 0.7902347 0.5153487 0.4021794 0.6751189
#> [22,] 0.4018763 0.3256960 0.4994346 0.6295773 0.2448955 0.2006715 0.4274084
#> [23,] 0.7542469 0.4651371 0.8211697 0.8173662 0.5048813 0.4099046 0.7544934
#> [24,] 0.6822751 0.4784419 0.7770891 0.7868451 0.4085806 0.4012465 0.6758047
#> [25,] 0.5181314 0.3934617 0.6178147 0.6889494 0.3243287 0.2814355 0.5363576
#> [26,] 0.6995727 0.5011176 0.7833861 0.8059429 0.4526205 0.4041440 0.6910062
#> [27,] 0.6730173 0.4736604 0.7435709 0.7664548 0.5146920 0.4409280 0.6975650
#> [28,] 0.6102117 0.3565542 0.6506639 0.6971862 0.3252510 0.2719737 0.5987506
#> [29,] 0.7040477 0.5303437 0.7742687 0.7815026 0.5345885 0.4544333 0.7061529
#> [30,] 0.5866505 0.4156007 0.6578481 0.6906649 0.4004834 0.3041416 0.5964140
#> [31,] 0.6867197 0.4693375 0.7505472 0.7676692 0.4802017 0.4125661 0.6725676
#> [32,] 0.7284960 0.5257559 0.7984065 0.8134990 0.5618236 0.4292245 0.7271705
#> [33,] 0.7243152 0.4860585 0.7979227 0.7943832 0.4851524 0.4037556 0.7097119
#> [34,] 0.5862297 0.3997999 0.6683803 0.6874384 0.4205179 0.3164912 0.5941929
#> [35,] 0.4921025 0.3267436 0.5377800 0.7057483 0.2192121 0.2842195 0.4812689
#> [36,] 0.7135969 0.4991955 0.7733209 0.7995153 0.5672687 0.4133250 0.7062904
#> [37,] 0.6876519 0.4135269 0.7724689 0.7822102 0.4379445 0.3787682 0.7089064
#> [38,] 0.7638281 0.5477506 0.8240720 0.8164028 0.5528603 0.4962326 0.7618191
#> [39,] 0.7127765 0.4570028 0.7917736 0.7875756 0.4275272 0.4007845 0.7107867
#> [40,] 0.7496779 0.4937633 0.8286617 0.8247338 0.5368167 0.4135409 0.7541444
#>            [,8]      [,9]     [,10]
#>  [1,] 0.9617851 0.7917960 0.6061281
#>  [2,] 0.9556033 0.7460586 0.5324480
#>  [3,] 0.9393208 0.6630132 0.5540417
#>  [4,] 0.9315540 0.6920324 0.4624997
#>  [5,] 0.9206599 0.6943267 0.4385209
#>  [6,] 0.9633462 0.7982129 0.6302611
#>  [7,] 0.9420719 0.6907558 0.6198954
#>  [8,] 0.9495928 0.6727977 0.5781452
#>  [9,] 0.9017567 0.5571948 0.4829357
#> [10,] 0.9617805 0.7964790 0.5785825
#> [11,] 0.9281629 0.6174373 0.4960090
#> [12,] 0.9327459 0.6636134 0.3285040
#> [13,] 0.9527288 0.7392934 0.5228441
#> [14,] 0.9418820 0.7056086 0.6180394
#> [15,] 0.9185807 0.6678047 0.4938657
#> [16,] 0.9613966 0.7716236 0.6893732
#> [17,] 0.9467030 0.6929674 0.5732486
#> [18,] 0.9320519 0.7248464 0.5268678
#> [19,] 0.9389217 0.6854673 0.5449630
#> [20,] 0.8684398 0.5344895 0.2630026
#> [21,] 0.9360822 0.7274018 0.4909722
#> [22,] 0.8176256 0.4841903 0.2043190
#> [23,] 0.9710945 0.7755255 0.6388714
#> [24,] 0.9500613 0.7023265 0.5788925
#> [25,] 0.8684218 0.5785759 0.3380552
#> [26,] 0.9494137 0.7015716 0.5702546
#> [27,] 0.9236976 0.7040735 0.5723637
#> [28,] 0.9383848 0.6470498 0.3794160
#> [29,] 0.9452987 0.7445898 0.6102317
#> [30,] 0.9142337 0.6165435 0.3760122
#> [31,] 0.9325664 0.7164716 0.5467710
#> [32,] 0.9560163 0.7728666 0.5668766
#> [33,] 0.9599568 0.7452917 0.6084615
#> [34,] 0.9146119 0.6595910 0.4607606
#> [35,] 0.8330020 0.5519110 0.2967024
#> [36,] 0.9469030 0.7575621 0.5408298
#> [37,] 0.9543256 0.7362413 0.5292766
#> [38,] 0.9665150 0.7836504 0.7016689
#> [39,] 0.9560231 0.7128116 0.6022658
#> [40,] 0.9716422 0.7929527 0.6178183
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5909103 0.4090897
#>  [2,] 0.5972487 0.4027513
#>  [3,] 0.5933185 0.4066815
#>  [4,] 0.5936638 0.4063362
#>  [5,] 0.5937834 0.4062166
#>  [6,] 0.5902320 0.4097680
#>  [7,] 0.5877359 0.4122641
#>  [8,] 0.5940050 0.4059950
#>  [9,] 0.5905701 0.4094299
#> [10,] 0.5896370 0.4103630
#> [11,] 0.5950919 0.4049081
#> [12,] 0.5993120 0.4006880
#> [13,] 0.5972371 0.4027629
#> [14,] 0.5878971 0.4121029
#> [15,] 0.5907537 0.4092463
#> [16,] 0.5894976 0.4105024
#> [17,] 0.5939117 0.4060883
#> [18,] 0.5890472 0.4109528
#> [19,] 0.5920203 0.4079797
#> [20,] 0.5983362 0.4016638
#> [21,] 0.5899632 0.4100368
#> [22,] 0.5884134 0.4115866
#> [23,] 0.5959352 0.4040648
#> [24,] 0.5926601 0.4073399
#> [25,] 0.5905655 0.4094345
#> [26,] 0.5924332 0.4075668
#> [27,] 0.5871874 0.4128126
#> [28,] 0.6006505 0.3993495
#> [29,] 0.5865810 0.4134190
#> [30,] 0.5935025 0.4064975
#> [31,] 0.5904776 0.4095224
#> [32,] 0.5914691 0.4085309
#> [33,] 0.5931100 0.4068900
#> [34,] 0.5903298 0.4096702
#> [35,] 0.5912692 0.4087308
#> [36,] 0.5912220 0.4087780
#> [37,] 0.5979559 0.4020441
#> [38,] 0.5875892 0.4124108
#> [39,] 0.5948919 0.4051081
#> [40,] 0.5949800 0.4050200
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]         [,2]         [,3]      [,4]        [,5]        [,6]
#>  [1,]  1.07746719  0.108674260  1.475472920 1.5774461  0.38081048 -0.22316614
#>  [2,]  0.79845373 -0.274741041  1.235829149 1.3003433 -0.17548733 -0.46908810
#>  [3,]  0.62411491 -0.222819376  1.067656742 1.1843105 -0.57385486 -0.48403970
#>  [4,]  0.71165158 -0.187822364  0.994611312 1.1807185 -0.03618313 -0.54653312
#>  [5,]  0.51337647 -0.146320483  0.913363054 1.2482370 -0.46225567 -0.57065381
#>  [6,]  1.11960616  0.046106302  1.486733188 1.5355746  0.42483663 -0.18934425
#>  [7,]  0.84304614  0.040200503  1.159170241 1.2398337 -0.26450659 -0.18412602
#>  [8,]  0.80752601 -0.184543472  1.092770081 1.1889231 -0.40411845 -0.48849682
#>  [9,]  0.36260950 -0.396201336  0.740602976 0.8667180 -0.89016129 -0.58842212
#> [10,]  0.98125431  0.234151195  1.441876013 1.5260262  0.40345606 -0.25247890
#> [11,]  0.43345386 -0.467932728  0.848647763 0.9628643 -0.70904820 -0.66693883
#> [12,]  0.31404315 -0.407162440  0.755212555 1.2018283 -0.52845901 -1.00003378
#> [13,]  0.69051002 -0.368493422  1.101604329 1.1845534 -0.27242009 -0.57418355
#> [14,]  0.83822657  0.004221051  1.208465344 1.1756168 -0.13502056 -0.22544474
#> [15,]  0.45722545 -0.250154404  0.929888819 1.1153944 -0.32756421 -0.49266358
#> [16,]  1.11980271  0.068748463  1.496851330 1.4114947  0.07762449 -0.08937429
#> [17,]  0.79393546 -0.162473508  1.200947291 1.2429288 -0.40818157 -0.43942711
#> [18,]  0.73719204  0.096473065  1.205785836 1.3366401  0.04050402 -0.34120729
#> [19,]  0.76070923 -0.079568168  1.168832140 1.3350216 -0.29771072 -0.40422722
#> [20,]  0.08952687 -0.844147942  0.285980276 0.9814317 -1.28528592 -1.07547640
#> [21,]  0.69770302  0.122189689  1.162976044 1.3263405  0.06141408 -0.39639245
#> [22,] -0.39765320 -0.727716559 -0.002261539 0.5304040 -1.12602429 -1.38210255
#> [23,]  1.12139234 -0.139678084  1.524292324 1.4986042  0.01952582 -0.36435968
#> [24,]  0.76424711 -0.086285728  1.248782992 1.3060121 -0.36983619 -0.40027390
#> [25,]  0.07255760 -0.432783947  0.480283032 0.7952124 -0.73394906 -0.93735213
#> [26,]  0.84526382  0.004470209  1.285508955 1.4238606 -0.19008842 -0.38822767
#> [27,]  0.72186371 -0.105456181  1.064612070 1.1884002  0.05878489 -0.23739682
#> [28,]  0.44820217 -0.590351452  0.621958705 0.8339343 -0.72974336 -0.98463186
#> [29,]  0.86664783  0.121524306  1.232573617 1.2744446  0.13857533 -0.18277386
#> [30,]  0.35013567 -0.340859365  0.653719088 0.8032297 -0.40345116 -0.82765277
#> [31,]  0.78482770 -0.122803992  1.101533007 1.1951965 -0.07923473 -0.35336719
#> [32,]  0.98700520  0.103114999  1.376364322 1.4729081  0.24856644 -0.28501584
#> [33,]  0.96596868 -0.055780580  1.373361409 1.3515518 -0.05940799 -0.38984078
#> [34,]  0.34840078 -0.406299117  0.700868315 0.7881705 -0.32064769 -0.76994401
#> [35,] -0.03159258 -0.722950623  0.151408673 0.8748232 -1.27026355 -0.92362664
#> [36,]  0.91291802 -0.003218058  1.227158800 1.3832677  0.27071609 -0.35023676
#> [37,]  0.78916468 -0.349404319  1.222304981 1.2785933 -0.24950844 -0.49477978
#> [38,]  1.17378259  0.191586351  1.544182867 1.4921635  0.21223447 -0.01507001
#> [39,]  0.90890728 -0.172414571  1.335649750 1.3103735 -0.29194744 -0.40219729
#> [40,]  1.09689516 -0.024948109  1.576172295 1.5487546  0.14753377 -0.34934661
#>              [,7]     [,8]        [,9]       [,10]
#>  [1,]  1.09862080 3.225565  1.33578531  0.43106562
#>  [2,]  0.94781189 3.069177  1.07770049  0.12997484
#>  [3,]  0.67281890 2.739556  0.67675130  0.21701461
#>  [4,]  0.67790849 2.610809  0.80963804 -0.15028337
#>  [5,]  0.53276523 2.451348  0.82042550 -0.24716699
#>  [6,]  1.07629338 3.268895  1.37516239  0.53333710
#>  [7,]  0.74358981 2.788879  0.80365529  0.48910426
#>  [8,]  0.76724914 2.935899  0.72086600  0.31516381
#>  [9,]  0.33857123 2.216898  0.22978505 -0.06828378
#> [10,]  1.01846099 3.225441  1.36443168  0.31695709
#> [11,]  0.60440710 2.558806  0.47868502 -0.01596443
#> [12,]  0.32399823 2.629656  0.67943844 -0.71495920
#> [13,]  0.83536036 3.003429  1.04229925  0.09144024
#> [14,]  0.79316198 2.785405  0.87415053  0.48123478
#> [15,]  0.62794760 2.423217  0.69827256 -0.02453849
#> [16,]  1.16389656 3.215046  1.21750195  0.79719041
#> [17,]  0.75654598 2.877106  0.81402890  0.29511778
#> [18,]  0.75894381 2.618645  0.96863037  0.10757473
#> [19,]  0.66008874 2.732575  0.77901274  0.18033925
#> [20,] -0.01951318 1.887234  0.13817733 -1.03042024
#> [21,]  0.73142944 2.684106  0.98148010 -0.03611512
#> [22,] -0.29243277 1.500343 -0.06325971 -1.35951564
#> [23,]  1.12272250 3.514391  1.23977850  0.57046896
#> [24,]  0.73455819 2.945730  0.85840113  0.31822867
#> [25,]  0.14568738 1.887076  0.31693020 -0.67197317
#> [26,]  0.80482777 2.932165  0.85479305  0.28288993
#> [27,]  0.83572917 2.493681  0.86677165  0.29150138
#> [28,]  0.40026187 2.723251  0.60609619 -0.49202760
#> [29,]  0.87677220 2.849614  1.06996233  0.44828633
#> [30,]  0.39054562 2.366460  0.47490263 -0.50650926
#> [31,]  0.71982081 2.626798  0.92702620  0.18763243
#> [32,]  0.98031367 3.078956  1.22456877  0.26911898
#> [33,]  0.89398535 3.176930  1.07365667  0.44084967
#> [34,]  0.38132598 2.371293  0.66147233 -0.15728105
#> [35,] -0.07495965 1.607054  0.20839491 -0.86305070
#> [36,]  0.87743511 2.881075  1.13935984  0.16368385
#> [37,]  0.89007865 3.039466  1.02652343  0.11724037
#> [38,]  1.16267876 3.362600  1.28706720  0.85525760
#> [39,]  0.89920793 3.079116  0.90907885  0.41491504
#> [40,]  1.12083957 3.534087  1.34281648  0.48029825
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1184817 -0.2492481
#>  [2,] 0.1397066 -0.2543078
#>  [3,] 0.1326676 -0.2450334
#>  [4,] 0.1309604 -0.2481719
#>  [5,] 0.1365164 -0.2431117
#>  [6,] 0.1166824 -0.2482422
#>  [7,] 0.1155824 -0.2390311
#>  [8,] 0.1329123 -0.2476347
#>  [9,] 0.1285993 -0.2377235
#> [10,] 0.1163444 -0.2461204
#> [11,] 0.1399136 -0.2451423
#> [12,] 0.1537433 -0.2488560
#> [13,] 0.1430049 -0.2509614
#> [14,] 0.1150722 -0.2402065
#> [15,] 0.1270353 -0.2400465
#> [16,] 0.1158411 -0.2460477
#> [17,] 0.1316600 -0.2485000
#> [18,] 0.1174580 -0.2425701
#> [19,] 0.1261926 -0.2461310
#> [20,] 0.1577783 -0.2407591
#> [21,] 0.1204820 -0.2433311
#> [22,] 0.1387564 -0.2186538
#> [23,] 0.1325170 -0.2560397
#> [24,] 0.1284014 -0.2465718
#> [25,] 0.1345365 -0.2317669
#> [26,] 0.1255470 -0.2484865
#> [27,] 0.1136855 -0.2386646
#> [28,] 0.1594100 -0.2487663
#> [29,] 0.1113212 -0.2385278
#> [30,] 0.1372666 -0.2411970
#> [31,] 0.1228191 -0.2431208
#> [32,] 0.1212163 -0.2488255
#> [33,] 0.1271607 -0.2496762
#> [34,] 0.1311862 -0.2341426
#> [35,] 0.1440947 -0.2251202
#> [36,] 0.1219909 -0.2470286
#> [37,] 0.1417203 -0.2552350
#> [38,] 0.1102046 -0.2438037
#> [39,] 0.1319451 -0.2522808
#> [40,] 0.1294220 -0.2551693
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5909103  0.5909103
#>  [2,]  0.4027513 -0.4027513
#>  [3,]  0.4066815 -0.4066815
#>  [4,]  0.4063362 -0.4063362
#>  [5,] -0.5937834  0.5937834
#>  [6,]  0.4097680 -0.4097680
#>  [7,]  0.4122641 -0.4122641
#>  [8,] -0.5940050  0.5940050
#>  [9,]  0.4094299 -0.4094299
#> [10,] -0.5896370  0.5896370
#> [11,]  0.4049081 -0.4049081
#> [12,]  0.4006880 -0.4006880
#> [13,]  0.4027629 -0.4027629
#> [14,] -0.5878971  0.5878971
#> [15,] -0.5907537  0.5907537
#> [16,] -0.5894976  0.5894976
#> [17,]  0.4060883 -0.4060883
#> [18,] -0.5890472  0.5890472
#> [19,] -0.5920203  0.5920203
#> [20,]  0.4016638 -0.4016638
#> [21,] -0.5899632  0.5899632
#> [22,] -0.5884134  0.5884134
#> [23,]  0.4040648 -0.4040648
#> [24,]  0.4073399 -0.4073399
#> [25,]  0.4094345 -0.4094345
#> [26,]  0.4075668 -0.4075668
#> [27,] -0.5871874  0.5871874
#> [28,]  0.3993495 -0.3993495
#> [29,] -0.5865810  0.5865810
#> [30,]  0.4064975 -0.4064975
#> [31,]  0.4095224 -0.4095224
#> [32,]  0.4085309 -0.4085309
#> [33,]  0.4068900 -0.4068900
#> [34,]  0.4096702 -0.4096702
#> [35,]  0.4087308 -0.4087308
#> [36,]  0.4087780 -0.4087780
#> [37,]  0.4020441 -0.4020441
#> [38,] -0.5875892  0.5875892
#> [39,]  0.4051081 -0.4051081
#> [40,]  0.4050200 -0.4050200
#> 
#> $L
#> [1] 0.6899233 0.6392527 0.8136660 0.6977174 0.6834472 0.6507256
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5797101