Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11          V12         V13         V14
#>  [1,] -0.3375407  0.004512865 -0.02068557 -0.001701359 -0.01382238 -0.04354302
#>  [2,] -0.2995811 -0.151626341 -0.03188639 -0.005555209 -0.09760174  0.03521585
#>  [3,] -0.2693688 -0.129953416  0.03011425 -0.034008573 -0.03953386 -0.03687422
#>  [4,] -0.3791920 -0.142713704 -0.09543951 -0.053155256 -0.00611431 -0.08514124
#>  [5,] -0.2739717  0.009141156 -0.08373602 -0.092022076 -0.04666300 -0.09325131
#>  [6,] -0.2507356 -0.078397743 -0.06238094 -0.008299963  0.07908034 -0.08838821
#>  [7,] -0.4113943 -0.060610234 -0.10231971 -0.070207995 -0.10132931 -0.17286402
#>  [8,] -0.2823984 -0.147853992 -0.07225386 -0.091863368 -0.10742295 -0.02542219
#>  [9,] -0.1963754 -0.179838930 -0.12882796 -0.103642734 -0.12425068 -0.09086708
#> [10,] -0.3045688 -0.110861668 -0.19379451 -0.031541759 -0.12080840 -0.03698081
#>                V15         V16         V17          V18         V19         V2
#>  [1,]  0.042174352  0.07405712  0.08962700  0.062494173  0.12226756 -0.2416066
#>  [2,]  0.028851711 -0.01530131 -0.01938018 -0.074609069 -0.01335272 -0.3485817
#>  [3,] -0.043872345 -0.07342333 -0.12540134 -0.095069814 -0.05261145 -0.3453596
#>  [4,] -0.073841902 -0.13548399 -0.10739863  0.087462612  0.05378014 -0.2144758
#>  [5,]  0.044913465  0.10640004 -0.05074760 -0.026245162  0.01723255 -0.2568479
#>  [6,] -0.002600078 -0.07485469  0.06742801 -0.021382628  0.07183977 -0.2915288
#>  [7,] -0.210806073 -0.13225326 -0.15400373 -0.064744501  0.02498626 -0.2215934
#>  [8,] -0.123889046 -0.03636441 -0.05787017  0.065920902 -0.01116459 -0.3082997
#>  [9,] -0.049608250  0.00256552  0.03295408  0.007027589  0.05514782 -0.3716553
#> [10,] -0.075254721 -0.01140942  0.06794184  0.055977637 -0.08460124 -0.3262444
#>                V20         V21          V22         V23          V24        V25
#>  [1,] -0.017451631  0.03995317  0.155262426  0.12618629  0.154326790 0.01982637
#>  [2,]  0.148251249  0.11595512  0.116341619  0.03848181  0.174139729 0.05604350
#>  [3,]  0.120690348  0.02436140  0.159982270  0.19911273  0.169525801 0.23955543
#>  [4,]  0.101824398 -0.01208546  0.165368275  0.03185954  0.120596222 0.11010604
#>  [5,]  0.009633224  0.07796501 -0.001421784  0.13850614  0.162787016 0.02500876
#>  [6,]  0.031762269  0.04315215  0.049759566  0.17381821 -0.007987137 0.10785164
#>  [7,]  0.115422796  0.11130661 -0.026374526  0.11957041  0.007868464 0.15607011
#>  [8,]  0.122912453  0.07455550  0.194293480  0.08966109  0.077233099 0.14213519
#>  [9,] -0.073576224  0.03823280 -0.001417097 -0.04168377  0.074661808 0.11434780
#> [10,] -0.062813230  0.10240717  0.092157531  0.08429012  0.137859838 0.09322380
#>              V26         V27         V28          V29         V3         V30
#>  [1,] 0.07457967 0.118340431 0.209326484  0.100979024 -0.2004994  0.02158750
#>  [2,] 0.11721546 0.004410469 0.002893387  0.007564055 -0.2391216  0.12264645
#>  [3,] 0.21477350 0.223360019 0.158189183  0.155124430 -0.3530186  0.12450439
#>  [4,] 0.06039784 0.123116675 0.060363197 -0.009246759 -0.3175851  0.08513272
#>  [5,] 0.17964661 0.068477214 0.110590716  0.180138942 -0.2914726  0.15338339
#>  [6,] 0.09924465 0.050061553 0.128037292  0.065879626 -0.1977734  0.03955635
#>  [7,] 0.20572582 0.224543248 0.205324996  0.171699673 -0.4124012  0.04481080
#>  [8,] 0.13262102 0.190159913 0.259049907  0.209886341 -0.2995688  0.12311464
#>  [9,] 0.11076483 0.193171351 0.006842733  0.057038201 -0.2399299  0.01152893
#> [10,] 0.07317770 0.080048939 0.115438865 -0.024287551 -0.1828056 -0.11870234
#>                 V31          V32          V33          V34         V35
#>  [1,]  0.0854859982  0.112703853  0.123006172  0.090973063  0.07743677
#>  [2,]  0.1157602302 -0.040080411 -0.003977660  0.011587689  0.09005818
#>  [3,]  0.0671274613  0.115959920  0.098161384  0.016087940 -0.14999929
#>  [4,] -0.0473812787 -0.040586040  0.056962161 -0.031253045 -0.14113924
#>  [5,]  0.0364053564 -0.059362598 -0.083427401 -0.078002378 -0.04619948
#>  [6,] -0.0003201617  0.098103848 -0.034629591 -0.027396551  0.07607525
#>  [7,]  0.0879949521  0.089325199  0.018432482  0.055250371 -0.18816423
#>  [8,]  0.1588726508  0.051092504  0.105626826 -0.005198546 -0.07820392
#>  [9,]  0.0622098192 -0.008072488 -0.008441841 -0.079145527 -0.05737445
#> [10,] -0.0321301133 -0.035880904 -0.136908297 -0.157441741 -0.01356521
#>               V36         V37          V38         V39         V4          V40
#>  [1,]  0.07272716  0.09312515  0.017357117  0.01479093 -0.2310332  0.047256665
#>  [2,]  0.05025408  0.10304721 -0.007806624  0.02562982 -0.3138145  0.029536669
#>  [3,] -0.03167013 -0.08712356 -0.119228493  0.03976608 -0.3782331 -0.041738948
#>  [4,] -0.13818081 -0.15831227 -0.043373435 -0.10006290 -0.1666094 -0.008300552
#>  [5,] -0.10119682 -0.12952192 -0.075857726 -0.12643134 -0.2618985 -0.023853908
#>  [6,] -0.04992722  0.06313907  0.005758604  0.04545049 -0.2892343 -0.042611229
#>  [7,] -0.08153667 -0.04687442 -0.018837955  0.03394762 -0.2657670 -0.120781271
#>  [8,] -0.06963201 -0.06807960  0.050422989  0.07337164 -0.2282820 -0.029655724
#>  [9,] -0.01141047 -0.10147918  0.038503689 -0.13011005 -0.1873156  0.005812159
#> [10,] -0.08536132 -0.10656182 -0.131888374 -0.14243065 -0.2023784 -0.174013515
#>               V41         V42         V43          V44         V45         V46
#>  [1,] -0.01191826 -0.04316354 -0.08905873 -0.072273137 -0.11669148 -0.08450182
#>  [2,] -0.02783186 -0.05424383 -0.11327842 -0.091272926 -0.14934213 -0.21636347
#>  [3,]  0.01889776 -0.05490750 -0.01389395 -0.095084581 -0.06126827 -0.15472919
#>  [4,] -0.11700042 -0.03640963 -0.17546782 -0.138637307 -0.21724316 -0.07137795
#>  [5,]  0.03857037 -0.10033414 -0.13090214 -0.108874011 -0.02033356 -0.15304035
#>  [6,] -0.02150301 -0.09022810 -0.01233157 -0.001586583 -0.16414359 -0.19872294
#>  [7,] -0.16943671 -0.03178206 -0.04192617 -0.014282840 -0.09067520 -0.08814950
#>  [8,]  0.07170103 -0.01008589  0.01207343 -0.026076676 -0.12411096 -0.20625580
#>  [9,] -0.14326840 -0.06933831 -0.13212392 -0.003570814 -0.08072679 -0.11010694
#> [10,] -0.17585059 -0.06509864 -0.12174660 -0.243566575 -0.05666014 -0.17091993
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.2055415 -0.1314596 -0.3434577 -0.1802405 -0.2770036 -0.3917449
#>  [2,] -0.2640856 -0.2629322 -0.1680373 -0.2261411 -0.2890890 -0.3541379
#>  [3,] -0.1898057 -0.2716224 -0.2950721 -0.2363637 -0.3106029 -0.3312245
#>  [4,] -0.1959115 -0.1771803 -0.3472193 -0.2398176 -0.3133266 -0.3322840
#>  [5,] -0.1476295 -0.2906408 -0.3502746 -0.3089082 -0.3001687 -0.2305887
#>  [6,] -0.1177800 -0.2716681 -0.2230467 -0.2701899 -0.2819390 -0.3318458
#>  [7,] -0.2329770 -0.2192547 -0.2041475 -0.2782096 -0.2515954 -0.3338420
#>  [8,] -0.1279355 -0.1913518 -0.2057129 -0.3091844 -0.3564921 -0.2456743
#>  [9,] -0.2178028 -0.1388558 -0.2934678 -0.2096442 -0.2805076 -0.3034531
#> [10,] -0.1199887 -0.2048903 -0.3341385 -0.1223780 -0.3742147 -0.2549047
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3400868 -0.3764824 -0.3345875 -0.3863678 -0.2732964 -0.3606429
#>  [2,] -0.3144258 -0.2494567 -0.3724697 -0.3155930 -0.3848651 -0.3633172
#>  [3,] -0.3563371 -0.3246566 -0.3360120 -0.2811955 -0.2632976 -0.3500976
#>  [4,] -0.3290692 -0.2967486 -0.3565428 -0.2920848 -0.3409851 -0.3419841
#>  [5,] -0.2422601 -0.2915639 -0.3031947 -0.2902817 -0.2611898 -0.3030828
#>  [6,] -0.2116685 -0.3489570 -0.3921187 -0.4017098 -0.3587371 -0.3444238
#>  [7,] -0.4339014 -0.3360679 -0.4234549 -0.3251579 -0.3551227 -0.3449902
#>  [8,] -0.3602658 -0.2840886 -0.3342344 -0.3860115 -0.3365755 -0.2721609
#>  [9,] -0.2913265 -0.3664918 -0.2468392 -0.2968369 -0.3762878 -0.2542704
#> [10,] -0.2735913 -0.2363452 -0.2545736 -0.3948335 -0.3530445 -0.2760078
#>              V58        V59         V6        V60         V7          V8
#>  [1,] -0.4343251 -0.2445194 -0.1777897 -0.2984907 -0.1536106 -0.11294056
#>  [2,] -0.3353700 -0.3251377 -0.1726492 -0.4420324 -0.1861777 -0.09358827
#>  [3,] -0.2946511 -0.4437020 -0.2736691 -0.3439959 -0.1231751 -0.21033397
#>  [4,] -0.2708351 -0.2531478 -0.1272173 -0.3424090 -0.1301406 -0.12785700
#>  [5,] -0.3687784 -0.3242667 -0.2548451 -0.2864817 -0.1995149 -0.11873630
#>  [6,] -0.2826555 -0.3452566 -0.1030403 -0.2525844 -0.1925635 -0.07828630
#>  [7,] -0.3407746 -0.3357051 -0.2519987 -0.2771392 -0.2147594 -0.24137508
#>  [8,] -0.3214447 -0.3948938 -0.1909430 -0.3587040 -0.2303272 -0.24766230
#>  [9,] -0.3061621 -0.3849555 -0.1592029 -0.3532393 -0.1680029 -0.26884180
#> [10,] -0.2549606 -0.2819606 -0.1165446 -0.2335109 -0.1835271 -0.09991949
#>                V9
#>  [1,] -0.11900571
#>  [2,] -0.04253033
#>  [3,] -0.05607134
#>  [4,] -0.08228114
#>  [5,] -0.09084774
#>  [6,] -0.10306772
#>  [7,] -0.19663507
#>  [8,] -0.22473926
#>  [9,] -0.22219089
#> [10,] -0.06983818
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.05278971 -0.01443138  0.04300603 -0.04420206 -0.03681650 -0.04832914
#> [2,] 0.03438922 -0.08305286 -0.13532301  0.03708421 -0.04336988 -0.06791488
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] 0.085179202 -0.04244999 -0.01076550 0.041996441
#> [2,] 0.003853575 -0.04351271 -0.01516908 0.005914118
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10           V11           V12          V13
#>  [1,] 2.050458e-04 0.0011941745  0.0013003282  0.0012962907 0.0014600838
#>  [2,] 1.404113e-04 0.0008473709  0.0007352384  0.0007229608 0.0011435833
#>  [3,] 3.060313e-05 0.0002040188 -0.0001952181 -0.0002605506 0.0003521224
#>  [4,] 3.620357e-04 0.0022882355  0.0028690065  0.0029806664 0.0028317211
#>  [5,] 2.566348e-04 0.0015643203  0.0017334407  0.0017424702 0.0019461239
#>  [6,] 2.338685e-04 0.0014218760  0.0015272828  0.0015334468 0.0017806140
#>  [7,] 1.142343e-04 0.0006970025  0.0005124551  0.0004849477 0.0009699472
#>  [8,] 2.582905e-04 0.0014890713  0.0016441581  0.0016236897 0.0017891097
#>  [9,] 1.468402e-04 0.0009289078  0.0009364067  0.0009802277 0.0012843612
#> [10,] 1.217569e-04 0.0007739031  0.0007355885  0.0007571829 0.0010895542
#>                V14         V15         V16         V17         V18         V19
#>  [1,] 0.0016150483 0.001810089 0.002020441 0.002165392 0.002346936 0.002780205
#>  [2,] 0.0015248417 0.001932739 0.002335007 0.002537259 0.002868143 0.002920804
#>  [3,] 0.0009353826 0.001538000 0.002045017 0.002231535 0.002673234 0.002151661
#>  [4,] 0.0026232875 0.002509315 0.002591934 0.002863652 0.002963744 0.004201190
#>  [5,] 0.0021113405 0.002337368 0.002625000 0.002837695 0.003069961 0.003647013
#>  [6,] 0.0019973260 0.002266738 0.002580614 0.002788495 0.003040738 0.003511752
#>  [7,] 0.0014300217 0.001927168 0.002388711 0.002598673 0.002931712 0.002801695
#>  [8,] 0.0019610511 0.002192469 0.002442325 0.002600562 0.002789556 0.003342297
#>  [9,] 0.0015589434 0.001842559 0.002184710 0.002396182 0.002656111 0.002858032
#> [10,] 0.0013697923 0.001668986 0.002012158 0.002203428 0.002468613 0.002585430
#>                 V2         V20         V21         V22         V23         V24
#>  [1,] 2.533152e-04 0.003257168 0.003463254 0.003192957 0.003019623 0.003129785
#>  [2,] 1.891456e-04 0.002969823 0.003018065 0.003087729 0.003364802 0.003804616
#>  [3,] 7.142913e-05 0.001644490 0.001502910 0.002010511 0.002801242 0.003643168
#>  [4,] 4.295022e-04 0.005440550 0.005995127 0.005282215 0.004389678 0.003944637
#>  [5,] 3.172174e-04 0.004261612 0.004535034 0.004210213 0.003954057 0.004048041
#>  [6,] 2.927181e-04 0.004012262 0.004240366 0.003996810 0.003859539 0.004030349
#>  [7,] 1.632015e-04 0.002691009 0.002684512 0.002873058 0.003273549 0.003796390
#>  [8,] 3.148314e-04 0.003972492 0.004226490 0.003836010 0.003544870 0.003606166
#>  [9,] 1.922379e-04 0.003033100 0.003129255 0.003120386 0.003243929 0.003490748
#> [10,] 1.657830e-04 0.002684424 0.002756692 0.002777126 0.002940101 0.003249617
#>               V25         V26         V27         V28         V29           V3
#>  [1,] 0.003012884 0.003102506 0.003120921 0.003206503 0.003006561 0.0002610388
#>  [2,] 0.003684008 0.003785391 0.004014671 0.003978814 0.003731899 0.0002273649
#>  [3,] 0.003516216 0.003556420 0.004040433 0.003869133 0.003644417 0.0001374444
#>  [4,] 0.003734198 0.003790329 0.003479811 0.003655621 0.003329020 0.0003986055
#>  [5,] 0.003896027 0.003992817 0.003980851 0.004065930 0.003791437 0.0003236068
#>  [6,] 0.003881840 0.003983935 0.004037846 0.004108968 0.003838765 0.0003072149
#>  [7,] 0.003663944 0.003734035 0.004039509 0.003955756 0.003689868 0.0002115383
#>  [8,] 0.003442528 0.003517200 0.003476442 0.003587045 0.003364735 0.0003153646
#>  [9,] 0.003373137 0.003489705 0.003663338 0.003638432 0.003375287 0.0002272716
#> [10,] 0.003147753 0.003234130 0.003414712 0.003392057 0.003154588 0.0001993640
#>               V30         V31         V32         V33         V34         V35
#>  [1,] 0.002728727 0.002157316 0.001916118 0.001949926 0.001933368 0.001812287
#>  [2,] 0.003483814 0.003119571 0.002495761 0.002736710 0.002958561 0.003155069
#>  [3,] 0.003570025 0.003649533 0.002604504 0.003025630 0.003555570 0.004187077
#>  [4,] 0.002702314 0.001556835 0.001844794 0.001518374 0.001019114 0.000260459
#>  [5,] 0.003421066 0.002679310 0.002418166 0.002446283 0.002397497 0.002207764
#>  [6,] 0.003484221 0.002812978 0.002459441 0.002525469 0.002527656 0.002419233
#>  [7,] 0.003522038 0.003277767 0.002519678 0.002796219 0.003115804 0.003472601
#>  [8,] 0.003072608 0.002366180 0.002163848 0.002198901 0.002179959 0.002023025
#>  [9,] 0.003085724 0.002682471 0.002230829 0.002327946 0.002423133 0.002485273
#> [10,] 0.002900277 0.002542526 0.002074083 0.002208133 0.002328990 0.002424702
#>                V36          V37         V38         V39           V4
#>  [1,] 1.643658e-03 0.0015623623 0.001702982 0.001643636 0.0003525014
#>  [2,] 2.979562e-03 0.0024995632 0.002096453 0.001672971 0.0002799967
#>  [3,] 4.169548e-03 0.0032687815 0.002180311 0.001392698 0.0001225736
#>  [4,] 4.818035e-05 0.0006287867 0.001724431 0.002278283 0.0005845677
#>  [5,] 1.995901e-03 0.0019270464 0.002161601 0.002121662 0.0004410969
#>  [6,] 2.210966e-03 0.0020588517 0.002154695 0.002031927 0.0004103792
#>  [7,] 3.348008e-03 0.0027625733 0.002168338 0.001632992 0.0002421071
#>  [8,] 1.822663e-03 0.0017742444 0.002026966 0.001985346 0.0004300140
#>  [9,] 2.343915e-03 0.0020618046 0.001876139 0.001623892 0.0002823378
#> [10,] 2.281868e-03 0.0019729576 0.001737435 0.001451218 0.0002428981
#>               V40         V41         V42          V43          V44
#>  [1,] 0.001557769 0.001477926 0.001452621 0.0012776163 0.0012076789
#>  [2,] 0.002024081 0.001843026 0.001404514 0.0010697630 0.0008980452
#>  [3,] 0.002237630 0.001907108 0.001027277 0.0005405489 0.0002841633
#>  [4,] 0.001528003 0.001584170 0.002227570 0.0022268858 0.0022940270
#>  [5,] 0.001967393 0.001873026 0.001877306 0.0016675335 0.0015864578
#>  [6,] 0.001993051 0.001880767 0.001788463 0.0015453464 0.0014459063
#>  [7,] 0.002099923 0.001879704 0.001317311 0.0009337840 0.0007336276
#>  [8,] 0.001784928 0.001690578 0.001719446 0.0015364356 0.0014710930
#>  [9,] 0.001768889 0.001629856 0.001371839 0.0011151282 0.0009773516
#> [10,] 0.001639613 0.001502483 0.001202795 0.0009528987 0.0008184444
#>                V45          V46          V47          V48           V49
#>  [1,] 1.269857e-03 0.0010288534 0.0006730606 4.814597e-04  2.681899e-04
#>  [2,] 8.443126e-04 0.0007355476 0.0004960406 3.197870e-04  1.702129e-04
#>  [3,] 9.789938e-05 0.0001819079 0.0001499184 3.409902e-05 -2.350215e-06
#>  [4,] 2.482476e-03 0.0018735818 0.0012412636 9.449235e-04  5.474635e-04
#>  [5,] 1.658014e-03 0.0013150280 0.0008596125 6.212031e-04  3.507439e-04
#>  [6,] 1.492801e-03 0.0012042979 0.0007954859 5.664540e-04  3.162368e-04
#>  [7,] 6.503138e-04 0.0005921413 0.0004129576 2.542304e-04  1.306091e-04
#>  [8,] 1.565073e-03 0.0012461436 0.0008007715 5.810675e-04  3.271826e-04
#>  [9,] 9.442847e-04 0.0007843698 0.0005445879 3.800742e-04  2.110100e-04
#> [10,] 7.750768e-04 0.0006500290 0.0004522520 3.087324e-04  1.702087e-04
#>                 V5          V50          V51          V52          V53
#>  [1,] 0.0004567362 1.053412e-04 9.311014e-05 8.639449e-05 6.031468e-05
#>  [2,] 0.0003585569 9.546083e-05 6.626851e-05 7.388161e-05 5.214616e-05
#>  [3,] 0.0001232989 6.060568e-05 1.415389e-05 3.599862e-05 2.501999e-05
#>  [4,] 0.0008299339 1.646134e-04 1.741484e-04 1.382944e-04 9.665090e-05
#>  [5,] 0.0005913675 1.351254e-04 1.201994e-04 1.093419e-04 7.628801e-05
#>  [6,] 0.0005450810 1.280188e-04 1.102366e-04 1.028005e-04 7.202098e-05
#>  [7,] 0.0003023349 8.969606e-05 5.568286e-05 6.522360e-05 4.516369e-05
#>  [8,] 0.0005580727 1.265558e-04 1.126833e-04 1.036684e-04 7.272074e-05
#>  [9,] 0.0003775592 9.705851e-05 7.587205e-05 7.308499e-05 5.193350e-05
#> [10,] 0.0003227283 8.373176e-05 6.264009e-05 6.367858e-05 4.492268e-05
#>                V54          V55          V56          V57          V58
#>  [1,] 6.757116e-05 5.317036e-05 3.559697e-05 3.959721e-05 4.332063e-05
#>  [2,] 7.144707e-05 5.168311e-05 2.985959e-05 4.478101e-05 3.342069e-05
#>  [3,] 6.118300e-05 3.790245e-05 1.142858e-05 4.061898e-05 1.157282e-05
#>  [4,] 8.951160e-05 7.590578e-05 6.052930e-05 4.462234e-05 7.587911e-05
#>  [5,] 8.613621e-05 6.788425e-05 4.640469e-05 4.913076e-05 5.580476e-05
#>  [6,] 8.324101e-05 6.458170e-05 4.337092e-05 4.885384e-05 5.136721e-05
#>  [7,] 7.049870e-05 4.919830e-05 2.657537e-05 4.533172e-05 2.860264e-05
#>  [8,] 8.202299e-05 6.551669e-05 4.431474e-05 4.646098e-05 5.413144e-05
#>  [9,] 6.550978e-05 4.834281e-05 3.312671e-05 4.215334e-05 3.605910e-05
#> [10,] 5.997813e-05 4.401774e-05 2.824170e-05 3.786110e-05 3.056142e-05
#>                V59           V6          V60           V7           V8
#>  [1,] 4.331273e-05 0.0005639241 3.563436e-05 0.0006153566 0.0006834716
#>  [2,] 4.317749e-05 0.0006237806 3.505583e-05 0.0006481804 0.0005289741
#>  [3,] 3.113151e-05 0.0005577165 2.449032e-05 0.0005312291 0.0001970528
#>  [4,] 5.985873e-05 0.0007551265 4.956405e-05 0.0008626347 0.0012240284
#>  [5,] 5.507211e-05 0.0007232348 4.520905e-05 0.0007912573 0.0008863422
#>  [6,] 5.275220e-05 0.0007090021 4.335343e-05 0.0007688296 0.0008193335
#>  [7,] 4.155510e-05 0.0006243752 3.356667e-05 0.0006388310 0.0004581205
#>  [8,] 5.278540e-05 0.0006623643 4.365589e-05 0.0007357105 0.0008471829
#>  [9,] 4.158196e-05 0.0005869409 3.355832e-05 0.0006272681 0.0005743451
#> [10,] 3.733831e-05 0.0005289632 3.013040e-05 0.0005589671 0.0004826038
#>                 V9
#>  [1,] 1.018640e-03
#>  [2,] 6.529657e-04
#>  [3,] 1.037835e-05
#>  [4,] 2.020860e-03
#>  [5,] 1.328140e-03
#>  [6,] 1.195042e-03
#>  [7,] 4.964093e-04
#>  [8,] 1.278624e-03
#>  [9,] 7.547326e-04
#> [10,] 6.121965e-04
#> 
#> $vW[[2]]
#>              [,1]        [,2]         [,3]         [,4]         [,5]
#> [1,]  0.008635813  0.01248664  0.007710805  0.004304299  0.007142734
#> [2,] -0.008635813 -0.01248664 -0.007710805 -0.004304299 -0.007142734
#>              [,6]        [,7]         [,8]       [,9]        [,10]
#> [1,]  0.008265917  0.01117566  0.008703213  0.0180486  0.008631728
#> [2,] -0.008265917 -0.01117566 -0.008703213 -0.0180486 -0.008631728
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.7574222 0.5440581 0.4822164 0.3268515 0.5030491 0.4956562 0.4414221
#>  [8] 0.5362346 0.3515554 0.3569813
#> 
#> $B[[2]]
#> [1]  0.1613742 -0.1254268
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.004793098 0.005808008 0.005567495 0.005521106 0.006099120 0.006092498
#>  [7] 0.005825680 0.005543564 0.005300311 0.004897536
#> 
#> $vB[[2]]
#> [1]  0.002090849 -0.002090849
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#>  [2,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#>  [3,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#>  [4,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [5,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#>  [6,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#>  [7,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [8,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#>  [9,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [10,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [11,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [12,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [13,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [14,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [15,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [16,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [17,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#> [18,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [19,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [20,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [21,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [22,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [23,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [24,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [25,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [26,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [27,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [28,] 0.0126 0.4284 0.3015 0.1207 0.3299 0.5707 0.6962 0.9751 1.0000 0.9293
#> [29,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [30,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [31,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [32,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [33,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#> [34,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [35,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [36,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [37,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [38,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [39,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [40,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#>  [2,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#>  [3,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#>  [4,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [5,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#>  [6,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#>  [7,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [8,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#>  [9,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [10,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [11,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [12,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [13,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [14,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [15,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [16,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [17,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#> [18,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [19,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [20,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [21,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [22,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [23,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [24,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [25,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [26,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [27,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [28,] 0.6210 0.0519 0.4586 0.5001 0.5032 0.7082 0.8420 0.8109 0.7690 0.8105
#> [29,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [30,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [31,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [32,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [33,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#> [34,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [35,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [36,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [37,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [38,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [39,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [40,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#>  [2,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#>  [3,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#>  [4,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [5,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#>  [6,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#>  [7,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [8,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#>  [9,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [10,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [11,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [12,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [13,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [14,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [15,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [16,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [17,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#> [18,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [19,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [20,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [21,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [22,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [23,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [24,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [25,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [26,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [27,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [28,] 0.6203 0.2356 0.0621 0.2595 0.6299 0.6762 0.2903 0.4393 0.8529 0.7180
#> [29,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [30,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [31,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [32,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [33,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#> [34,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [35,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [36,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [37,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [38,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [39,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [40,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#>  [2,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#>  [3,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#>  [4,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [5,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#>  [6,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#>  [7,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [8,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#>  [9,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [10,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [11,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [12,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [13,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [14,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [15,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [16,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [17,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#> [18,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [19,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [20,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [21,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [22,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [23,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [24,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [25,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [26,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [27,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [28,] 0.4801 0.5856 0.4993 0.0518 0.2866 0.0601 0.1167 0.2737 0.2812 0.2078
#> [29,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [30,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [31,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [32,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [33,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#> [34,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [35,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [36,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [37,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [38,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [39,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [40,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#>  [2,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#>  [3,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#>  [4,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [5,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#>  [6,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#>  [7,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [8,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#>  [9,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [10,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [11,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [12,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [13,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [14,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [15,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [16,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [17,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#> [18,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [19,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [20,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [21,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [22,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [23,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [24,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [25,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [26,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [27,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [28,] 0.0660 0.0491 0.0345 0.0172 0.1072 0.0287 0.0027 0.0208 0.0048 0.0199
#> [29,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [30,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [31,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [32,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [33,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#> [34,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [35,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [36,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [37,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [38,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [39,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [40,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#>  [2,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#>  [3,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#>  [4,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [5,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#>  [6,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#>  [7,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [8,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#>  [9,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [10,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [11,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [12,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [13,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [14,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [15,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [16,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [17,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#> [18,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [19,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [20,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [21,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [22,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [23,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [24,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [25,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [26,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [27,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [28,] 0.0126 0.0022 0.0037 0.0034 0.0114 0.2587 0.0077 0.2304 0.2067 0.3416
#> [29,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [30,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [31,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [32,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [33,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#> [34,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [35,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [36,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [37,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [38,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [39,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [40,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8374121 0.6858632 0.8095499 0.5379321 0.6597627 0.6942164 0.6970378
#>  [2,] 0.8074365 0.6322789 0.7732056 0.4687127 0.6054286 0.6461527 0.6533955
#>  [3,] 0.8582617 0.7225306 0.7531545 0.5930582 0.7120698 0.7173535 0.6347481
#>  [4,] 0.8405479 0.7030712 0.7274554 0.4891161 0.6605147 0.6692818 0.5427388
#>  [5,] 0.8297533 0.6976904 0.8158314 0.5926540 0.7039504 0.7054570 0.7396052
#>  [6,] 0.8785238 0.7382952 0.8057790 0.5472233 0.6848568 0.7349528 0.7182200
#>  [7,] 0.7595612 0.6444012 0.5501257 0.3413912 0.4701503 0.5673086 0.4417781
#>  [8,] 0.8116226 0.6983082 0.6826346 0.4608113 0.5662741 0.6509805 0.5914915
#>  [9,] 0.8398782 0.7038142 0.8142858 0.5686445 0.6673356 0.7086684 0.6979302
#> [10,] 0.7811395 0.6294250 0.5529209 0.4633111 0.5419856 0.6181249 0.4447654
#> [11,] 0.8725667 0.7461035 0.8480981 0.5825317 0.7005899 0.7322371 0.7699480
#> [12,] 0.8179168 0.7070440 0.7033520 0.5270366 0.6346170 0.6740948 0.6620643
#> [13,] 0.8794054 0.7584980 0.8205132 0.6404797 0.7386407 0.7470255 0.7100993
#> [14,] 0.8702217 0.7513925 0.8105551 0.5857411 0.6512869 0.7213655 0.6727972
#> [15,] 0.8414260 0.6940040 0.7687767 0.5987408 0.7033411 0.7079342 0.6912210
#> [16,] 0.8427176 0.7304736 0.7711046 0.5586937 0.6611176 0.7155885 0.6822391
#> [17,] 0.8228432 0.7047347 0.8038969 0.5880784 0.7002996 0.6964515 0.7205233
#> [18,] 0.8567936 0.7036248 0.7525656 0.5602420 0.6781602 0.7156632 0.6718552
#> [19,] 0.8335382 0.7071758 0.7527392 0.5026139 0.6328015 0.7010040 0.6990619
#> [20,] 0.8487987 0.7332891 0.8182791 0.6113879 0.7117907 0.7211468 0.6992693
#> [21,] 0.8276787 0.6712550 0.7106814 0.4516005 0.5983748 0.6423279 0.5724767
#> [22,] 0.8160531 0.6881894 0.7212519 0.5188626 0.6580620 0.6608011 0.5213833
#> [23,] 0.8556371 0.7196299 0.7941840 0.5509408 0.6304440 0.7039204 0.6639502
#> [24,] 0.7461680 0.5972455 0.5287948 0.4507202 0.5401271 0.5963446 0.4228898
#> [25,] 0.7688456 0.5833439 0.6746044 0.3975674 0.5821776 0.5994736 0.5544404
#> [26,] 0.8356751 0.7281023 0.7348652 0.5719419 0.6521964 0.6972375 0.6202638
#> [27,] 0.7754247 0.6135449 0.5694003 0.4652103 0.5553424 0.6129939 0.4183794
#> [28,] 0.8570276 0.6925945 0.6839260 0.4457013 0.6027783 0.6870485 0.5252511
#> [29,] 0.8117983 0.6650346 0.6717126 0.4545460 0.6125779 0.6293301 0.4880214
#> [30,] 0.8318445 0.7173900 0.8372649 0.6336668 0.7266827 0.7227413 0.7634123
#> [31,] 0.8101457 0.6840870 0.6254191 0.4075367 0.5176368 0.6431693 0.5593877
#> [32,] 0.8144616 0.6419975 0.7342014 0.4461399 0.6175677 0.6359667 0.6071744
#> [33,] 0.8283188 0.7101958 0.7779897 0.6043094 0.6628331 0.7060582 0.7078535
#> [34,] 0.8252659 0.7425055 0.7476315 0.5548765 0.6501106 0.6869018 0.6423223
#> [35,] 0.8488534 0.7112008 0.8518112 0.6266105 0.7333357 0.7295832 0.7716451
#> [36,] 0.8585917 0.7386766 0.8228519 0.5843207 0.6971738 0.7270468 0.7521376
#> [37,] 0.8126259 0.6562917 0.7639880 0.5031019 0.6214929 0.6696399 0.6491504
#> [38,] 0.8672782 0.7561278 0.8140495 0.5889267 0.7052706 0.7360950 0.7431974
#> [39,] 0.8332140 0.6771032 0.8001804 0.5090850 0.6368866 0.6819962 0.6831705
#> [40,] 0.7753812 0.6256987 0.7026218 0.4519785 0.5927884 0.6228785 0.6183952
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8133041 0.5021143 0.4301046
#>  [2,] 0.7923976 0.4735612 0.3812220
#>  [3,] 0.8038864 0.5563859 0.5272384
#>  [4,] 0.7411224 0.5159513 0.4871528
#>  [5,] 0.8228575 0.5880073 0.5237903
#>  [6,] 0.8349782 0.6044409 0.5049092
#>  [7,] 0.6109149 0.3269154 0.2651828
#>  [8,] 0.7218254 0.4995480 0.3777983
#>  [9,] 0.7915308 0.5429952 0.5023997
#> [10,] 0.6212255 0.4523684 0.4362954
#> [11,] 0.8566880 0.6009585 0.4645408
#> [12,] 0.7699956 0.5459199 0.4379252
#> [13,] 0.8336394 0.6312107 0.6181451
#> [14,] 0.8160839 0.5322572 0.4648543
#> [15,] 0.7884289 0.6279143 0.5829614
#> [16,] 0.7955637 0.5937232 0.5158824
#> [17,] 0.8087295 0.5696556 0.4949114
#> [18,] 0.8147613 0.5500797 0.4802508
#> [19,] 0.8030741 0.5480997 0.4448161
#> [20,] 0.8297637 0.5682846 0.5389160
#> [21,] 0.7467038 0.4642876 0.3506550
#> [22,] 0.7201057 0.4727666 0.4861539
#> [23,] 0.7781349 0.4735051 0.4443982
#> [24,] 0.5910157 0.4455690 0.4380634
#> [25,] 0.7206598 0.4282044 0.3567787
#> [26,] 0.7603995 0.5428549 0.5302811
#> [27,] 0.6110211 0.4286862 0.4584812
#> [28,] 0.7140175 0.5086675 0.4675058
#> [29,] 0.6900295 0.4789959 0.4693182
#> [30,] 0.8316252 0.6128020 0.5802641
#> [31,] 0.7100141 0.4754545 0.3384806
#> [32,] 0.7758354 0.4736019 0.3985229
#> [33,] 0.8032425 0.5571102 0.5248202
#> [34,] 0.7833448 0.5415172 0.4503831
#> [35,] 0.8433307 0.6184924 0.5730125
#> [36,] 0.8398147 0.6160324 0.4913630
#> [37,] 0.7667030 0.4588054 0.4237835
#> [38,] 0.8443664 0.5932089 0.5021206
#> [39,] 0.8091909 0.4953629 0.4149637
#> [40,] 0.7527856 0.4709232 0.3442200
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6619941 0.3380059
#>  [2,] 0.6575860 0.3424140
#>  [3,] 0.6601829 0.3398171
#>  [4,] 0.6563220 0.3436780
#>  [5,] 0.6647780 0.3352220
#>  [6,] 0.6659418 0.3340582
#>  [7,] 0.6398989 0.3601011
#>  [8,] 0.6534903 0.3465097
#>  [9,] 0.6634104 0.3365896
#> [10,] 0.6419909 0.3580091
#> [11,] 0.6681296 0.3318704
#> [12,] 0.6570581 0.3429419
#> [13,] 0.6672340 0.3327660
#> [14,] 0.6632104 0.3367896
#> [15,] 0.6624809 0.3375191
#> [16,] 0.6625826 0.3374174
#> [17,] 0.6633299 0.3366701
#> [18,] 0.6603999 0.3396001
#> [19,] 0.6610465 0.3389535
#> [20,] 0.6647352 0.3352648
#> [21,] 0.6535159 0.3464841
#> [22,] 0.6539740 0.3460260
#> [23,] 0.6609084 0.3390916
#> [24,] 0.6392625 0.3607375
#> [25,] 0.6492557 0.3507443
#> [26,] 0.6586202 0.3413798
#> [27,] 0.6417054 0.3582946
#> [28,] 0.6540878 0.3459122
#> [29,] 0.6508020 0.3491980
#> [30,] 0.6671883 0.3328117
#> [31,] 0.6497417 0.3502583
#> [32,] 0.6555465 0.3444535
#> [33,] 0.6619318 0.3380682
#> [34,] 0.6592199 0.3407801
#> [35,] 0.6682218 0.3317782
#> [36,] 0.6666667 0.3333333
#> [37,] 0.6576059 0.3423941
#> [38,] 0.6664900 0.3335100
#> [39,] 0.6610534 0.3389466
#> [40,] 0.6526688 0.3473312
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]      [,2]      [,3]        [,4]       [,5]      [,6]        [,7]
#>  [1,] 1.639098 0.7808498 1.4470881  0.15202036  0.6622371 0.8199060  0.83323179
#>  [2,] 1.433439 0.5420059 1.2265010 -0.12531310  0.4281366 0.6021703  0.63399822
#>  [3,] 1.800927 0.9570491 1.1155078  0.37662219  0.9054577 0.9313718  0.55264097
#>  [4,] 1.662310 0.8619658 0.9817505 -0.04354229  0.6655887 0.7049384  0.17137347
#>  [5,] 1.583880 0.8363238 1.4883564  0.37494780  0.8661810 0.8734208  1.04391740
#>  [6,] 1.978524 1.0371265 1.4228129  0.18945797  0.7761826 1.0198983  0.93564952
#>  [7,] 1.150275 0.5945184 0.2011787 -0.65710080 -0.1195408 0.2708785 -0.23394882
#>  [8,] 1.460588 0.8392548 0.7659062 -0.15707679  0.2666656 0.6233520  0.37013457
#>  [9,] 1.657322 0.8655275 1.4781022  0.27632270  0.6961591 0.8889254  0.83746082
#> [10,] 1.272319 0.5297508 0.2124795 -0.14702002  0.1683390 0.4815968 -0.22184389
#> [11,] 1.923846 1.0779376 1.7197615  0.33317513  0.8501083 1.0060024  1.20801735
#> [12,] 1.502297 0.8810705 0.8633113  0.10825194  0.5520756 0.7267643  0.67250707
#> [13,] 1.986811 1.1444626 1.5198286  0.57744697  1.0389154 1.0828109  0.89586636
#> [14,] 1.902921 1.1060528 1.4536208  0.34638694  0.6247007 0.9512452  0.72086384
#> [15,] 1.668877 0.8189055 1.2014168  0.40022100  0.8632588 0.8853722  0.80583365
#> [16,] 1.678589 0.9970268 1.2145588  0.23586235  0.6682787 0.9226831  0.76408086
#> [17,] 1.535730 0.8699473 1.4108306  0.35602729  0.8487248 0.8304566  0.94705862
#> [18,] 1.788910 0.8646191 1.1123429  0.24214442  0.7453295 0.9230503  0.71658766
#> [19,] 1.610914 0.8817067 1.1132751  0.01045566  0.5442542 0.8520835  0.84283451
#> [20,] 1.725210 1.0113748 1.5047312  0.45314993  0.9040971 0.9501572  0.84382073
#> [21,] 1.569264 0.7138665 0.8986956 -0.19420620  0.3986980 0.5854821  0.29196329
#> [22,] 1.489832 0.7916679 0.9506801  0.07548637  0.6546695 0.6668662  0.08558549
#> [23,] 1.779516 0.9426268 1.3503326  0.20447263  0.5341222 0.8660368  0.68094764
#> [24,] 1.078278 0.3940011 0.1153067 -0.19776133  0.1608542 0.3902572 -0.31092170
#> [25,] 1.201804 0.3365158 0.7290849 -0.41561127  0.3317191 0.4032722  0.21862824
#> [26,] 1.626394 0.9850158 1.0194488  0.28977831  0.6287079 0.8341775  0.49066791
#> [27,] 1.239200 0.4622379 0.2794047 -0.13938396  0.2222804 0.4599144 -0.32942978
#> [28,] 1.790819 0.8122769 0.7718732 -0.21805460  0.4170549 0.7863567  0.10109032
#> [29,] 1.461738 0.6858119 0.7159410 -0.18231947  0.4581612 0.5293438 -0.04792356
#> [30,] 1.598757 0.9315516 1.6380166  0.54797978  0.9778564 0.9581006  1.17147919
#> [31,] 1.450957 0.7726180 0.5126141 -0.37415803  0.0705764 0.5891465  0.23867762
#> [32,] 1.479265 0.5840445 1.0160447 -0.21627954  0.4792370 0.5579008  0.43545046
#> [33,] 1.573759 0.8963352 1.2539896  0.42345382  0.6759452 0.8763161  0.88498159
#> [34,] 1.552440 1.0590317 1.0860197  0.22039382  0.6195255 0.7856744  0.58545771
#> [35,] 1.725636 0.9012231 1.7488775  0.51770300  1.0116129 0.9925088  1.21762401
#> [36,] 1.803642 1.0391017 1.5357899  0.34053596  0.8338757 0.9796904  1.11004535
#> [37,] 1.467164 0.6468121 1.1746695  0.01240760  0.4958897 0.7065570  0.61530688
#> [38,] 1.877105 1.1315663 1.4765404  0.35953034  0.8725239 1.0257698  1.06265381
#> [39,] 1.608579 0.7404911 1.3874221  0.03634414  0.5618765 0.7629606  0.76838046
#> [40,] 1.238950 0.5138080 0.8598142 -0.19268011  0.3755045 0.5017842  0.48274241
#>            [,8]         [,9]        [,10]
#>  [1,] 1.4716238  0.008457202 -0.281424210
#>  [2,] 1.3394386 -0.105853766 -0.484364490
#>  [3,] 1.4107642  0.226507193  0.109061665
#>  [4,] 1.0518106  0.063826888 -0.051400308
#>  [5,] 1.5358288  0.355733868  0.095233234
#>  [6,] 1.6213284  0.424003578  0.019637616
#>  [7,] 0.4511597 -0.722169572 -1.019202338
#>  [8,] 0.9535345 -0.001807886 -0.498904030
#>  [9,] 1.3341774  0.172406590  0.009598908
#> [10,] 0.4947529 -0.191106092 -0.256210996
#> [11,] 1.7880497  0.409460333 -0.142075112
#> [12,] 1.2082862  0.184198479 -0.249586636
#> [13,] 1.6116436  0.537414367  0.481682423
#> [14,] 1.4900376  0.129208437 -0.140814934
#> [15,] 1.3154809  0.523279374  0.334942342
#> [16,] 1.3587944  0.379378369  0.063551067
#> [17,] 1.4417756  0.280445981 -0.020355235
#> [18,] 1.4812498  0.200992680 -0.079037947
#> [19,] 1.4056194  0.192995556 -0.221638560
#> [20,] 1.5839536  0.274855578  0.155979308
#> [21,] 1.0811090 -0.143093446 -0.616161228
#> [22,] 0.9449862 -0.109041672 -0.055398443
#> [23,] 1.2548306 -0.106078784 -0.223330683
#> [24,] 0.3681659 -0.218590440 -0.249025232
#> [25,] 0.9477368 -0.289180908 -0.589372919
#> [26,] 1.1548711  0.171841024  0.121272708
#> [27,] 0.4516063 -0.287213460 -0.166458508
#> [28,] 0.9149769  0.034673300 -0.130160296
#> [29,] 0.8002572 -0.084065945 -0.122881629
#> [30,] 1.5971897  0.459105678  0.323857595
#> [31,] 0.8954525 -0.098260928 -0.670072330
#> [32,] 1.2415601 -0.105690834 -0.411623564
#> [33,] 1.4066845  0.229441998  0.099362465
#> [34,] 1.2852659  0.166451870 -0.199122905
#> [35,] 1.6832218  0.483153921  0.294152968
#> [36,] 1.6568501  0.472741408 -0.034551424
#> [37,] 1.1897872 -0.165152716 -0.307260678
#> [38,] 1.6910819  0.377246627  0.008482648
#> [39,] 1.4447614 -0.018548985 -0.343482773
#> [40,] 1.1135244 -0.116438540 -0.644544581
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2509240 -0.4212695
#>  [2,] 0.2474608 -0.4050940
#>  [3,] 0.2442236 -0.4198859
#>  [4,] 0.2404013 -0.4065450
#>  [5,] 0.2559522 -0.4287080
#>  [6,] 0.2562508 -0.4336362
#>  [7,] 0.2242266 -0.3506987
#>  [8,] 0.2407125 -0.3937045
#>  [9,] 0.2540304 -0.4244992
#> [10,] 0.2262078 -0.3578078
#> [11,] 0.2584316 -0.4413061
#> [12,] 0.2450929 -0.4051183
#> [13,] 0.2564031 -0.4392983
#> [14,] 0.2496686 -0.4279652
#> [15,] 0.2537520 -0.4206176
#> [16,] 0.2516712 -0.4231534
#> [17,] 0.2526028 -0.4255660
#> [18,] 0.2473071 -0.4177697
#> [19,] 0.2513205 -0.4166407
#> [20,] 0.2520708 -0.4323972
#> [21,] 0.2386327 -0.3958971
#> [22,] 0.2361232 -0.4004306
#> [23,] 0.2496747 -0.4176701
#> [24,] 0.2232804 -0.3488841
#> [25,] 0.2369017 -0.3788674
#> [26,] 0.2454502 -0.4117014
#> [27,] 0.2252096 -0.3575640
#> [28,] 0.2401419 -0.3969149
#> [29,] 0.2355644 -0.3870019
#> [30,] 0.2591638 -0.4363318
#> [31,] 0.2367686 -0.3811356
#> [32,] 0.2438253 -0.3996849
#> [33,] 0.2520811 -0.4198337
#> [34,] 0.2438834 -0.4159362
#> [35,] 0.2608409 -0.4393125
#> [36,] 0.2568662 -0.4362809
#> [37,] 0.2469811 -0.4056623
#> [38,] 0.2554963 -0.4368560
#> [39,] 0.2505142 -0.4174779
#> [40,] 0.2399718 -0.3908192
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3380059 -0.3380059
#>  [2,]  0.3424140 -0.3424140
#>  [3,] -0.6601829  0.6601829
#>  [4,] -0.6563220  0.6563220
#>  [5,]  0.3352220 -0.3352220
#>  [6,] -0.6659418  0.6659418
#>  [7,]  0.3601011 -0.3601011
#>  [8,] -0.6534903  0.6534903
#>  [9,]  0.3365896 -0.3365896
#> [10,]  0.3580091 -0.3580091
#> [11,] -0.6681296  0.6681296
#> [12,] -0.6570581  0.6570581
#> [13,] -0.6672340  0.6672340
#> [14,]  0.3367896 -0.3367896
#> [15,] -0.6624809  0.6624809
#> [16,] -0.6625826  0.6625826
#> [17,] -0.6633299  0.6633299
#> [18,]  0.3396001 -0.3396001
#> [19,]  0.3389535 -0.3389535
#> [20,]  0.3352648 -0.3352648
#> [21,] -0.6535159  0.6535159
#> [22,]  0.3460260 -0.3460260
#> [23,]  0.3390916 -0.3390916
#> [24,]  0.3607375 -0.3607375
#> [25,]  0.3507443 -0.3507443
#> [26,]  0.3413798 -0.3413798
#> [27,]  0.3582946 -0.3582946
#> [28,] -0.6540878  0.6540878
#> [29,] -0.6508020  0.6508020
#> [30,]  0.3328117 -0.3328117
#> [31,] -0.6497417  0.6497417
#> [32,]  0.3444535 -0.3444535
#> [33,] -0.6619318  0.6619318
#> [34,] -0.6592199  0.6592199
#> [35,]  0.3317782 -0.3317782
#> [36,] -0.6666667  0.6666667
#> [37,]  0.3423941 -0.3423941
#> [38,] -0.6664900  0.6664900
#> [39,]  0.3389466 -0.3389466
#> [40,]  0.3473312 -0.3473312
#> 
#> $L
#> [1] 0.7017376 0.7268106 0.7134469 0.8066720 0.8954392 0.7168990
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536