Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10           V11          V12           V13
#>  [1,] -0.2895300 -0.158715172 -0.1127940422 -0.095131061 -0.1188108094
#>  [2,] -0.3376679 -0.122820522 -0.0467383577 -0.007531348 -0.0114343689
#>  [3,] -0.2651151 -0.167897247 -0.0787144551 -0.005809045 -0.1393912902
#>  [4,] -0.2787479 -0.143797447 -0.1701882960 -0.028955118 -0.2057671906
#>  [5,] -0.3153156 -0.124971743 -0.1453930337 -0.107176872 -0.1773341736
#>  [6,] -0.3565254  0.005736798 -0.0547480212 -0.079650651  0.0254915570
#>  [7,] -0.2578740 -0.154857427  0.0003179955 -0.134007625  0.0003693836
#>  [8,] -0.3422190 -0.048595555 -0.1538171542 -0.030657249 -0.0901363302
#>  [9,] -0.3882696 -0.173420607 -0.1386376215 -0.080073013 -0.1727100322
#> [10,] -0.2415743 -0.061830840 -0.1583350571 -0.186485430  0.0125828436
#>               V14         V15          V16         V17           V18
#>  [1,] -0.03031696  0.03068252 -0.044635581 -0.04681851 -0.0067243249
#>  [2,]  0.01997406 -0.08907958 -0.057040804 -0.03435319  0.0553684095
#>  [3,] -0.16534158 -0.05247779  0.004500689 -0.11412211 -0.0792573316
#>  [4,] -0.19253859 -0.19402742 -0.070460404 -0.17902544 -0.1676557682
#>  [5,] -0.01028767 -0.06946588  0.064134450  0.09473395  0.0787092289
#>  [6,]  0.01640354  0.03999219  0.045455544  0.03994378  0.1023920118
#>  [7,] -0.06206045 -0.07210695  0.003301298  0.01021649  0.1057808961
#>  [8,] -0.15158505 -0.07565423  0.017950710  0.02186984 -0.0303185720
#>  [9,] -0.05871500  0.02697392 -0.149266331 -0.08399627 -0.1112948525
#> [10,] -0.08879473 -0.04360539 -0.049110949  0.04468973  0.0008751735
#>               V19         V2         V20           V21          V22        V23
#>  [1,]  0.02018222 -0.2909362  0.09079393  0.1550114552  0.089791041 0.06992416
#>  [2,]  0.11573762 -0.2523684  0.04135831  0.1056670815  0.180414474 0.09913202
#>  [3,] -0.01487935 -0.2850029 -0.06963157  0.1318659328  0.124125450 0.08156526
#>  [4,] -0.08051968 -0.2208928 -0.04375177 -0.1098348325 -0.002487094 0.09507507
#>  [5,]  0.11751848 -0.3067471  0.13158009  0.0559787986  0.085748440 0.01853507
#>  [6,]  0.07683698 -0.2980645  0.13817346  0.0468570182  0.014755890 0.10581532
#>  [7,] -0.02795913 -0.3811665  0.13776290  0.0634554137  0.056103610 0.17959739
#>  [8,] -0.03045647 -0.2418292  0.05981355 -0.0002862293  0.122223036 0.15396482
#>  [9,] -0.05288019 -0.2908477 -0.04917819  0.1355251373  0.066647831 0.19135530
#> [10,] -0.07156214 -0.2416085  0.02247693  0.1276511241  0.168880330 0.03266687
#>              V24         V25        V26         V27         V28          V29
#>  [1,] 0.05618117  0.09254292 0.20121294  0.06573402  0.16538407  0.121990565
#>  [2,] 0.03836782 -0.01086476 0.07493991  0.08919571  0.16060804  0.133474665
#>  [3,] 0.17456909  0.09401751 0.05973903  0.09381662 -0.06229721 -0.040917618
#>  [4,] 0.11013462  0.05308782 0.06749913  0.11420047  0.11274394  0.171137191
#>  [5,] 0.05414620  0.05340597 0.08840430 -0.01665025 -0.02673006  0.001389273
#>  [6,] 0.10090986  0.27222202 0.20849284  0.19707088  0.20167655  0.201824957
#>  [7,] 0.12176508  0.15280742 0.08326930  0.17245398  0.18122238  0.127006433
#>  [8,] 0.09175554  0.14881357 0.15199038  0.10971730  0.17061575  0.192312188
#>  [9,] 0.23118556  0.23303074 0.25277255  0.27526186  0.23478345  0.136680251
#> [10,] 0.23440944  0.20976790 0.25637427  0.30415525  0.12439100  0.072869772
#>               V3         V30         V31          V32         V33          V34
#>  [1,] -0.3207678 -0.07133244 -0.03612198  0.082691874 -0.01927312  0.027824054
#>  [2,] -0.2372124  0.20101377  0.07548557  0.069982583  0.09304112  0.047288749
#>  [3,] -0.2487621  0.10886057 -0.11550373 -0.042958895 -0.10848408  0.009050882
#>  [4,] -0.1660631  0.11627905 -0.09914321 -0.029947241 -0.04433737 -0.002232171
#>  [5,] -0.3104868 -0.05104619 -0.01158619 -0.004524174 -0.11582288 -0.195471193
#>  [6,] -0.2443804  0.07110382  0.08971895 -0.068627689  0.06403314 -0.005560875
#>  [7,] -0.2189088  0.10411726  0.16040347  0.080701831  0.10434481  0.018368026
#>  [8,] -0.2936674  0.04930086 -0.03036050 -0.100711245 -0.04982944 -0.013547365
#>  [9,] -0.3987754  0.18689284  0.16672419  0.064171415  0.00400966  0.070068134
#> [10,] -0.3151274  0.17101553  0.11209030 -0.044026606  0.02002707  0.020009938
#>               V35          V36         V37         V38         V39         V4
#>  [1,] -0.06398034 -0.090276218 -0.11566750 -0.07604515 -0.05614094 -0.2216649
#>  [2,]  0.07859409 -0.072683361 -0.04149787 -0.03786132 -0.08496066 -0.2910738
#>  [3,] -0.03486732 -0.026103275 -0.12344468 -0.10389732 -0.01689950 -0.3282426
#>  [4,] -0.10275637 -0.009169869 -0.14078588 -0.05169926 -0.04924927 -0.3500447
#>  [5,] -0.11159039  0.007806868 -0.12464224  0.02629585  0.01267144 -0.1783498
#>  [6,]  0.01805864 -0.063743842  0.05228145  0.02038331 -0.03408786 -0.2807127
#>  [7,] -0.03745136  0.087449656  0.07729143 -0.08804358 -0.01976519 -0.2155911
#>  [8,] -0.11888973 -0.161018418  0.01335882 -0.01549003  0.00303790 -0.2289524
#>  [9,]  0.03561853  0.098793115  0.06030932  0.04739237 -0.07610984 -0.3583753
#> [10,]  0.06893979  0.014203950  0.01728351 -0.04327212  0.03100557 -0.3544948
#>               V40         V41          V42          V43         V44         V45
#>  [1,] -0.06044751 -0.02856107  0.013099633 -0.003748675 -0.19600152 -0.06343721
#>  [2,]  0.10255757 -0.11296884 -0.019672142 -0.101757253 -0.03022185 -0.06679288
#>  [3,] -0.11467590 -0.02217763 -0.117769821 -0.139145870 -0.16225552 -0.15211398
#>  [4,] -0.13142187 -0.09992395 -0.113685233 -0.191624870 -0.19748675 -0.17684650
#>  [5,] -0.14368710 -0.04540341 -0.058207629 -0.057095764 -0.09646370 -0.11802486
#>  [6,] -0.07668631 -0.05356427  0.005016081 -0.091438550 -0.17306430 -0.06567282
#>  [7,]  0.00269683 -0.13537136 -0.084821392 -0.068897189 -0.00572656 -0.11275334
#>  [8,] -0.12809740 -0.08441759 -0.157314496 -0.046667424 -0.13144646 -0.10087006
#>  [9,] -0.08827866 -0.01158762 -0.029474179 -0.061268005 -0.05099415 -0.12957673
#> [10,]  0.04681361 -0.05942954 -0.074453293  0.018374334 -0.11999485 -0.16699044
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.15344118 -0.11501156 -0.1257067 -0.2554850 -0.3002280 -0.3879528
#>  [2,] -0.07231462 -0.12393217 -0.1680387 -0.2307797 -0.2231044 -0.2912215
#>  [3,] -0.22905938 -0.24508921 -0.2116864 -0.3172862 -0.2574767 -0.3581334
#>  [4,] -0.14137483 -0.15353741 -0.1865416 -0.1794748 -0.2545779 -0.3066306
#>  [5,] -0.08683711 -0.15231491 -0.2733205 -0.3181973 -0.2388167 -0.2556867
#>  [6,] -0.24095159 -0.25323339 -0.1924712 -0.3687489 -0.2012305 -0.4214540
#>  [7,] -0.13284902 -0.09047846 -0.2828548 -0.2664088 -0.3263490 -0.2599572
#>  [8,] -0.22659912 -0.25950921 -0.1673406 -0.1656842 -0.1646964 -0.3598694
#>  [9,] -0.10633213 -0.27012542 -0.2054238 -0.3600842 -0.2530419 -0.3028459
#> [10,] -0.13917750 -0.11695000 -0.1519825 -0.2463709 -0.1610841 -0.2767505
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.2749410 -0.2758211 -0.2356534 -0.2771190 -0.2527705 -0.3541834
#>  [2,] -0.4017082 -0.3658440 -0.4049820 -0.4070710 -0.3320720 -0.2944131
#>  [3,] -0.2223786 -0.2406178 -0.2737702 -0.3716409 -0.3251364 -0.2282929
#>  [4,] -0.2979935 -0.3747887 -0.3701157 -0.2322192 -0.3448717 -0.3558410
#>  [5,] -0.1894110 -0.3693499 -0.3037245 -0.3008968 -0.2360167 -0.3716951
#>  [6,] -0.3864911 -0.4579870 -0.3165810 -0.4047463 -0.2988921 -0.3682132
#>  [7,] -0.3747245 -0.3355621 -0.3203523 -0.4035733 -0.3416668 -0.2539642
#>  [8,] -0.3485280 -0.3763995 -0.4026296 -0.3788801 -0.3270768 -0.4163751
#>  [9,] -0.3925351 -0.2682494 -0.3019151 -0.2821501 -0.4578176 -0.3996651
#> [10,] -0.3452489 -0.3255808 -0.3785210 -0.3196943 -0.3087640 -0.3073141
#>              V57        V58        V59         V6        V60         V7
#>  [1,] -0.3748741 -0.2501721 -0.3777049 -0.2262480 -0.3084435 -0.1985358
#>  [2,] -0.4339645 -0.3989300 -0.4059359 -0.2243703 -0.3897334 -0.1874644
#>  [3,] -0.2617360 -0.2421162 -0.3750968 -0.1215058 -0.2789713 -0.1882713
#>  [4,] -0.3381661 -0.2388001 -0.3633712 -0.1267341 -0.2921639 -0.2405028
#>  [5,] -0.3555822 -0.3342065 -0.2907419 -0.2109854 -0.2129792 -0.1364551
#>  [6,] -0.3850759 -0.4575459 -0.3677921 -0.1699216 -0.3528322 -0.2630895
#>  [7,] -0.2982995 -0.3955014 -0.3519497 -0.2138641 -0.3848995 -0.1398613
#>  [8,] -0.3657608 -0.3543202 -0.2432414 -0.1609687 -0.3593864 -0.1130949
#>  [9,] -0.3964610 -0.3200355 -0.3852899 -0.1556393 -0.3365460 -0.1562928
#> [10,] -0.2136432 -0.3979430 -0.3612562 -0.2127830 -0.3288927 -0.2390942
#>               V8          V9
#>  [1,] -0.2137263 -0.03490687
#>  [2,] -0.2290241 -0.04872638
#>  [3,] -0.1487885 -0.19066617
#>  [4,] -0.2055542 -0.11641017
#>  [5,] -0.1139182 -0.19073748
#>  [6,] -0.1040167 -0.11909664
#>  [7,] -0.1525962 -0.18154616
#>  [8,] -0.1008173 -0.03774869
#>  [9,] -0.2192970 -0.16400930
#> [10,] -0.1669734 -0.17380626
#> 
#> $W[[2]]
#>            [,1]       [,2]        [,3]       [,4]         [,5]        [,6]
#> [1,] 0.01624050 0.06553874 -0.07616737 0.01627028 -0.007368606 -0.06909266
#> [2,] 0.03671023 0.08618976  0.04075964 0.01582309 -0.004673251  0.07193384
#>            [,7]        [,8]        [,9]      [,10]
#> [1,] 0.04052711  0.01626001 -0.08624342 -0.1045041
#> [2,] 0.09571745 -0.02053620  0.02863060 -0.0316663
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -3.634362e-06  1.047663e-04  1.141008e-04  1.349975e-04  0.0002008717
#>  [2,] -4.379827e-06  7.050016e-05  7.310225e-05  8.763323e-05  0.0001528746
#>  [3,]  2.203188e-05  7.590258e-04  9.239375e-04  1.047789e-03  0.0008702484
#>  [4,] -1.591776e-05 -2.011222e-04 -2.703136e-04 -2.828562e-04 -0.0001005730
#>  [5,] -1.836920e-05 -2.240577e-04 -3.005393e-04 -3.256665e-04 -0.0001394517
#>  [6,]  2.110736e-05  6.761978e-04  7.973958e-04  8.804573e-04  0.0007264549
#>  [7,] -9.531163e-07  1.922226e-04  2.170290e-04  2.507087e-04  0.0002688464
#>  [8,] -1.947528e-05 -3.073054e-04 -3.881269e-04 -4.233240e-04 -0.0002056550
#>  [9,]  8.355192e-06  4.463869e-04  5.141905e-04  5.731916e-04  0.0004967256
#> [10,]  3.013848e-06  3.294991e-04  3.785782e-04  4.243188e-04  0.0004004769
#>                V14           V15           V16           V17           V18
#>  [1,] 0.0001682604  1.574074e-04  1.305169e-04  1.235190e-04  1.460869e-04
#>  [2,] 0.0001451797  1.593907e-04  1.477687e-04  1.467290e-04  1.645549e-04
#>  [3,] 0.0002551750 -3.017483e-04 -5.687604e-04 -5.528592e-04 -3.417727e-04
#>  [4,] 0.0001189819  3.410108e-04  4.239305e-04  4.086485e-04  3.456977e-04
#>  [5,] 0.0001085781  3.549123e-04  4.489051e-04  4.304330e-04  3.566593e-04
#>  [6,] 0.0002070497 -2.887209e-04 -5.600405e-04 -5.567909e-04 -4.087891e-04
#>  [7,] 0.0001412316  4.362076e-05 -3.144562e-05 -3.159005e-05  2.142109e-05
#>  [8,] 0.0001245718  4.671562e-04  6.040100e-04  5.860905e-04  4.960434e-04
#>  [9,] 0.0001079046 -2.458476e-04 -4.753695e-04 -4.683024e-04 -3.302877e-04
#> [10,] 0.0001503393 -6.831130e-05 -2.118717e-04 -2.141676e-04 -1.294278e-04
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0002744283 -9.937476e-06 0.0004574209 0.0005106742 0.0005098760
#>  [2,] 0.0002491438 -9.058414e-06 0.0003843830 0.0004244419 0.0004224966
#>  [3,] 0.0004583034  6.789803e-06 0.0010638472 0.0012262162 0.0011636732
#>  [4,] 0.0001882213 -1.973046e-05 0.0002033813 0.0002111148 0.0002141654
#>  [5,] 0.0001550671 -2.040836e-05 0.0001345504 0.0001368284 0.0001672486
#>  [6,] 0.0002840828  5.034268e-06 0.0008415137 0.0009405694 0.0008544593
#>  [7,] 0.0002422440 -9.373693e-06 0.0004704465 0.0005245098 0.0005019423
#>  [8,] 0.0001988759 -1.878750e-05 0.0001073168 0.0001061161 0.0001584185
#>  [9,] 0.0002074735 -1.053404e-05 0.0006536056 0.0007541138 0.0006599867
#> [10,] 0.0002435255 -9.408698e-06 0.0005842101 0.0006671378 0.0006201582
#>                V23          V24          V25          V26          V27
#>  [1,] 0.0005941939 0.0006903208 0.0008013818 0.0009438648 0.0009217888
#>  [2,] 0.0004964664 0.0005755834 0.0006673625 0.0007841141 0.0007588014
#>  [3,] 0.0011923318 0.0013352088 0.0015104832 0.0017238316 0.0019389531
#>  [4,] 0.0003330895 0.0004264189 0.0005143314 0.0006277320 0.0004960583
#>  [5,] 0.0002955683 0.0003871704 0.0004770082 0.0005862814 0.0004450080
#>  [6,] 0.0008438421 0.0009198846 0.0011044942 0.0012698452 0.0014553378
#>  [7,] 0.0005523353 0.0006322734 0.0007479899 0.0008804412 0.0009061643
#>  [8,] 0.0002844639 0.0003627979 0.0004272836 0.0005328423 0.0003616699
#>  [9,] 0.0006426413 0.0007293141 0.0008779525 0.0010261958 0.0011628512
#> [10,] 0.0006580473 0.0007585368 0.0009043859 0.0010624924 0.0011336706
#>                V28          V29            V3          V30          V31
#>  [1,] 0.0008360843 0.0006811998  3.731846e-05 0.0005484407 0.0004364024
#>  [2,] 0.0006930731 0.0005724153  2.840493e-05 0.0004647599 0.0003838073
#>  [3,] 0.0018478219 0.0012778600  1.327680e-04 0.0008058942 0.0003568473
#>  [4,] 0.0004026451 0.0004242724 -3.254711e-06 0.0004430510 0.0004812610
#>  [5,] 0.0003478251 0.0003944391 -1.024978e-05 0.0004339737 0.0004892401
#>  [6,] 0.0014018702 0.0009146672  1.155982e-04 0.0005831615 0.0002376098
#>  [7,] 0.0008480314 0.0006423513  4.555170e-05 0.0004766068 0.0003443697
#>  [8,] 0.0002591696 0.0003619104 -2.092478e-05 0.0004264528 0.0005144095
#>  [9,] 0.0011261262 0.0007172048  8.612650e-05 0.0004292990 0.0001898298
#> [10,] 0.0010674732 0.0007569757  6.711504e-05 0.0005247154 0.0003256855
#>                V32           V33           V34           V35           V36
#>  [1,] 0.0003584391  3.293575e-04  3.386668e-04  3.072178e-04  0.0002243920
#>  [2,] 0.0003119383  2.988751e-04  3.205730e-04  3.131137e-04  0.0002532353
#>  [3,] 0.0004000817 -2.101257e-05 -4.193120e-04 -9.481881e-04 -0.0013383256
#>  [4,] 0.0003479882  4.844875e-04  7.078603e-04  9.093197e-04  0.0009435248
#>  [5,] 0.0003452306  4.819776e-04  7.024456e-04  9.219100e-04  0.0009947608
#>  [6,] 0.0003155855 -5.572385e-05 -4.606397e-04 -9.979552e-04 -0.0014310333
#>  [7,] 0.0003049864  2.139774e-04  1.482166e-04  3.187573e-05 -0.0001128309
#>  [8,] 0.0003420011  5.484334e-04  8.261069e-04  1.132415e-03  0.0012753311
#>  [9,] 0.0002452676  1.529097e-05 -2.350888e-04 -5.676905e-04 -0.0009306080
#> [10,] 0.0003146689  1.461195e-04 -1.111524e-05 -2.386105e-04 -0.0004758099
#>                 V37           V38          V39            V4           V40
#>  [1,]  2.243468e-04  1.350484e-04 1.095799e-04  2.285525e-05  1.890789e-04
#>  [2,]  2.294292e-04  1.374910e-04 1.136002e-04  1.538425e-05  1.835992e-04
#>  [3,] -7.175763e-04 -1.096331e-04 1.639552e-04  1.569256e-04  4.945715e-05
#>  [4,]  6.652183e-04  2.723268e-04 9.887452e-05 -3.852712e-05  2.429204e-04
#>  [5,]  6.996946e-04  2.672261e-04 8.478083e-05 -4.903869e-05  2.458708e-04
#>  [6,] -8.494202e-04 -2.561843e-04 1.243191e-05  1.554419e-04 -7.299044e-05
#>  [7,] -6.856544e-06  5.032037e-05 9.006288e-05  4.271345e-05  1.265588e-04
#>  [8,]  8.857241e-04  3.318382e-04 1.012967e-04 -6.723226e-05  3.026264e-04
#>  [9,] -5.546914e-04 -1.465309e-04 4.086867e-05  1.108592e-04 -2.381597e-05
#> [10,] -2.347569e-04 -3.280634e-05 7.315833e-05  7.393363e-05  7.910938e-05
#>                 V41          V42           V43           V44           V45
#>  [1,]  1.545450e-04 1.463371e-04  1.223054e-04  7.374043e-05  9.516826e-05
#>  [2,]  1.512437e-04 1.296323e-04  9.540853e-05  4.809009e-05  5.514225e-05
#>  [3,] -3.820103e-05 2.736068e-04  5.091104e-04  4.960177e-04  7.749810e-04
#>  [4,]  2.237052e-04 8.499313e-05 -4.685592e-05 -1.036417e-04 -1.947109e-04
#>  [5,]  2.340815e-04 7.992104e-05 -6.472087e-05 -1.142324e-04 -2.178136e-04
#>  [6,] -9.921222e-05 1.487242e-04  3.624821e-04  3.684107e-04  7.173114e-04
#>  [7,]  8.778784e-05 1.284865e-04  1.507256e-04  1.122086e-04  1.905343e-04
#>  [8,]  2.867026e-04 8.442216e-05 -1.058737e-04 -1.770576e-04 -3.454939e-04
#>  [9,] -6.803712e-05 1.385717e-04  2.863145e-04  2.376695e-04  4.797815e-04
#> [10,]  3.685815e-05 1.477487e-04  2.251303e-04  1.901812e-04  3.445010e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.040635e-04  4.084602e-05  4.695132e-06 -1.171050e-05  4.732014e-05
#>  [2,]  6.984993e-05  2.854816e-05 -1.103133e-06 -1.239966e-05  3.543557e-05
#>  [3,]  6.234634e-04  3.238225e-04  2.382597e-04  1.157905e-04  2.304681e-04
#>  [4,] -1.197979e-04 -9.635438e-05 -1.082824e-04 -7.665219e-05 -3.134811e-05
#>  [5,] -1.354687e-04 -9.623666e-05 -1.087286e-04 -7.679878e-05 -4.415985e-05
#>  [6,]  5.789770e-04  2.972015e-04  2.272778e-04  1.211900e-04  1.865519e-04
#>  [7,]  1.723774e-04  8.109235e-05  4.244525e-05  1.288711e-05  6.456823e-05
#>  [8,] -2.288489e-04 -1.360617e-04 -1.444021e-04 -9.693145e-05 -6.390237e-05
#>  [9,]  3.830181e-04  1.744541e-04  1.280017e-04  6.518419e-05  1.338047e-04
#> [10,]  2.905081e-04  1.381800e-04  8.854650e-05  3.840288e-05  1.003979e-04
#>                 V50           V51           V52          V53          V54
#>  [1,]  1.084507e-06  1.150008e-05  7.473616e-06 9.882575e-06 7.750783e-06
#>  [2,]  1.483487e-06  8.823226e-06  5.227358e-06 8.447327e-06 6.355917e-06
#>  [3,]  8.022420e-06  4.942557e-05  4.672625e-05 1.501851e-05 1.472562e-05
#>  [4,] -2.889795e-06 -6.513172e-06 -9.476260e-06 8.513315e-06 4.937177e-06
#>  [5,] -3.600647e-06 -8.421808e-06 -1.233483e-05 6.886200e-06 3.495010e-06
#>  [6,]  1.328919e-05  4.215542e-05  4.293580e-05 1.062701e-05 1.095021e-05
#>  [7,]  3.392791e-06  1.589722e-05  1.298572e-05 9.048973e-06 7.289056e-06
#>  [8,] -4.948195e-06 -1.223447e-05 -1.797249e-05 6.746486e-06 3.914473e-06
#>  [9,]  6.678527e-06  3.238843e-05  3.165990e-05 1.174190e-05 1.085533e-05
#> [10,]  4.650026e-06  2.434217e-05  2.182745e-05 1.054172e-05 9.003243e-06
#>                 V55          V56           V57           V58          V59
#>  [1,]  5.612488e-06 4.585533e-06  6.688133e-06  5.937486e-06 1.406420e-06
#>  [2,]  5.272676e-06 3.875439e-06  6.118829e-06  5.054515e-06 1.331412e-06
#>  [3,] -9.744312e-07 5.901166e-06 -7.406797e-06  1.726976e-05 1.579099e-06
#>  [4,]  8.757458e-06 4.141877e-06  1.307960e-05  9.359644e-07 9.859788e-07
#>  [5,]  7.954300e-06 3.506938e-06  1.329032e-05 -2.164890e-07 9.320898e-08
#>  [6,] -7.620346e-07 4.639779e-06 -7.052750e-06  1.368330e-05 8.327240e-07
#>  [7,]  3.651519e-06 3.731779e-06  3.204792e-06  7.286139e-06 8.479912e-07
#>  [8,]  9.388621e-06 3.670158e-06  1.611119e-05 -1.151141e-06 1.080071e-06
#>  [9,]  7.391145e-07 4.175883e-06 -3.605987e-06  1.483517e-05 1.840903e-06
#> [10,]  2.681375e-06 4.193854e-06  5.608255e-07  1.045976e-05 1.057816e-06
#>                  V6          V60           V7            V8            V9
#>  [1,]  9.537530e-05 3.783001e-06 6.514812e-05  3.640506e-06  2.593368e-05
#>  [2,]  7.691701e-05 3.317472e-06 5.787943e-05  4.922046e-06  7.751849e-06
#>  [3,]  3.054860e-04 3.393744e-06 9.346743e-05  7.050075e-05  4.905793e-04
#>  [4,]  3.399662e-06 4.526119e-06 4.084977e-05 -4.835483e-05 -1.909995e-04
#>  [5,] -8.636666e-06 3.905136e-06 4.116022e-05 -4.299936e-05 -2.002365e-04
#>  [6,]  2.389567e-04 1.384273e-06 6.411822e-05  4.732586e-05  4.296787e-04
#>  [7,]  1.099619e-04 2.862119e-06 5.505116e-05  6.276858e-06  8.528450e-05
#>  [8,] -2.872508e-05 4.375988e-06 5.286200e-05 -3.438692e-05 -2.665682e-04
#>  [9,]  1.728295e-04 2.539046e-06 3.417676e-05 -1.506837e-06  2.509688e-04
#> [10,]  1.491461e-04 2.803855e-06 5.603528e-05  9.228244e-06  1.752684e-04
#> 
#> $vW[[2]]
#>              [,1]         [,2]        [,3]         [,4]         [,5]
#> [1,] -0.003377091 -0.001782529  0.00308696  0.001003899  0.005802665
#> [2,]  0.003377091  0.001782529 -0.00308696 -0.001003899 -0.005802665
#>              [,6]         [,7]         [,8]          [,9]         [,10]
#> [1,] -0.001580235  0.001079181 -0.004929348 -0.0004685366 -0.0006252486
#> [2,]  0.001580235 -0.001079181  0.004929348  0.0004685366  0.0006252486
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4490319 0.6997575 0.3623605 0.3247172 0.2120659 0.6557748 0.7015315
#>  [8] 0.4721977 0.6820738 0.4101229
#> 
#> $B[[2]]
#> [1] 0.008148695 0.022932537
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0009998501 0.0008752007 0.0010135620 0.0010222478 0.0009748779
#>  [6] 0.0006146996 0.0008014775 0.0010660298 0.0005840018 0.0008018534
#> 
#> $vB[[2]]
#> [1]  0.0001388494 -0.0001388494
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#>  [2,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#>  [3,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#>  [4,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>  [5,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#>  [6,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#>  [7,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [8,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#>  [9,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [10,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [11,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [12,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [13,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [14,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [15,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [16,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [17,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [18,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [19,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [20,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [21,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [22,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [23,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [24,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [25,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [26,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [27,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [28,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [29,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [30,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [31,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [32,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [33,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [34,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [35,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [36,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [37,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [38,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [39,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [40,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#>  [2,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#>  [3,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#>  [4,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>  [5,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#>  [6,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#>  [7,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [8,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#>  [9,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [10,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [11,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [12,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [13,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [14,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [15,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [16,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [17,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [18,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [19,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [20,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [21,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [22,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [23,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [24,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [25,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [26,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [27,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [28,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [29,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [30,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [31,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [32,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [33,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [34,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [35,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [36,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [37,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [38,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [39,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [40,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#>  [2,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#>  [3,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#>  [4,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>  [5,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#>  [6,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#>  [7,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [8,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#>  [9,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [10,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [11,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [12,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [13,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [14,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [15,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [16,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [17,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [18,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [19,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [20,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [21,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [22,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [23,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [24,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [25,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [26,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [27,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [28,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [29,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [30,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [31,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [32,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [33,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [34,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [35,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [36,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [37,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [38,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [39,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [40,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#>  [2,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#>  [3,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#>  [4,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>  [5,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#>  [6,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#>  [7,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [8,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#>  [9,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [10,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [11,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [12,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [13,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [14,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [15,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [16,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [17,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [18,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [19,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [20,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [21,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [22,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [23,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [24,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [25,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [26,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [27,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [28,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [29,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [30,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [31,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [32,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [33,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [34,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [35,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [36,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [37,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [38,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [39,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [40,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#>  [2,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#>  [3,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#>  [4,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>  [5,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#>  [6,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#>  [7,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [8,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#>  [9,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [10,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [11,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [12,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [13,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [14,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [15,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [16,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [17,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [18,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [19,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [20,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [21,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [22,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [23,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [24,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [25,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [26,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [27,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [28,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [29,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [30,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [31,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [32,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [33,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [34,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [35,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [36,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [37,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [38,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [39,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [40,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#>  [2,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#>  [3,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#>  [4,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#>  [5,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#>  [6,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#>  [7,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [8,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#>  [9,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [10,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [11,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [12,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [13,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [14,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [15,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [16,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [17,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [18,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [19,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [20,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [21,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [22,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [23,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [24,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [25,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [26,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [27,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [28,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [29,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [30,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [31,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [32,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [33,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [34,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [35,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [36,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [37,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [38,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [39,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [40,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6363187 0.7752863 0.4646879 0.3189461 0.5328885 0.8280392 0.7930447
#>  [2,] 0.6808173 0.7845967 0.5171899 0.5298378 0.4701536 0.8324756 0.8165879
#>  [3,] 0.5521772 0.6821536 0.4082107 0.2986400 0.4478468 0.7854161 0.7476562
#>  [4,] 0.6620338 0.7764174 0.4928508 0.3540776 0.5466711 0.8316308 0.8023196
#>  [5,] 0.4860037 0.6979682 0.3349505 0.2765087 0.3508922 0.6734331 0.7037287
#>  [6,] 0.6362054 0.7758224 0.5000358 0.4590714 0.4539061 0.8010602 0.7948907
#>  [7,] 0.6858693 0.7486155 0.5792156 0.5620543 0.5169523 0.8113011 0.7856479
#>  [8,] 0.6457388 0.8066237 0.4633732 0.4814984 0.4517445 0.8184601 0.8333758
#>  [9,] 0.6112661 0.7974566 0.4495904 0.4235726 0.4335275 0.8385479 0.8501586
#> [10,] 0.6595574 0.8102420 0.5009794 0.3896936 0.5517718 0.8440692 0.8183355
#> [11,] 0.5724238 0.7109693 0.4164100 0.3229565 0.4341952 0.7882943 0.7699329
#> [12,] 0.5543282 0.7538413 0.3996860 0.4210794 0.3525382 0.7420310 0.7682062
#> [13,] 0.6475102 0.7851936 0.4328933 0.3726620 0.4292487 0.8225684 0.7993189
#> [14,] 0.5563445 0.7241119 0.3297656 0.3172515 0.3406291 0.7227962 0.7246188
#> [15,] 0.6340695 0.7834259 0.4489025 0.3879118 0.4193714 0.8102087 0.7924376
#> [16,] 0.5170317 0.6659375 0.3663884 0.2445218 0.4597779 0.7444029 0.6982799
#> [17,] 0.6991219 0.8021683 0.5475732 0.4945980 0.5093087 0.8517650 0.8280568
#> [18,] 0.5878337 0.7517430 0.4586787 0.4792339 0.4500517 0.7508161 0.7574943
#> [19,] 0.6506412 0.7676536 0.4962918 0.3501673 0.5299085 0.8181333 0.7855320
#> [20,] 0.6448445 0.7667359 0.4605098 0.3330489 0.4422036 0.8094738 0.8081407
#> [21,] 0.6023084 0.7059247 0.4591845 0.3119795 0.4956429 0.7876355 0.7458471
#> [22,] 0.5927244 0.7452292 0.4411136 0.4089681 0.4136888 0.7831368 0.7759770
#> [23,] 0.6977882 0.7953077 0.5655101 0.5453409 0.5119723 0.8434422 0.8325897
#> [24,] 0.5744609 0.7615509 0.3836279 0.4216335 0.3397850 0.7923529 0.7987413
#> [25,] 0.5511263 0.6917175 0.3607192 0.2489068 0.4570567 0.7543004 0.6904749
#> [26,] 0.4819245 0.7139114 0.3048263 0.2644476 0.3590691 0.7176741 0.7268740
#> [27,] 0.7062042 0.7984841 0.5203531 0.4141828 0.5511021 0.8712425 0.8374494
#> [28,] 0.6564070 0.7759877 0.4810498 0.3993223 0.4651157 0.8286122 0.8137497
#> [29,] 0.6397068 0.7790350 0.4447973 0.3177357 0.5091205 0.8314316 0.7977698
#> [30,] 0.6786920 0.7812432 0.5103151 0.5169952 0.4748107 0.8331883 0.8103837
#> [31,] 0.6811163 0.8266747 0.5355048 0.5271021 0.4939465 0.8390283 0.8488147
#> [32,] 0.6244874 0.7527607 0.4431063 0.2856285 0.4797830 0.8030685 0.7870148
#> [33,] 0.6708668 0.8105067 0.5170206 0.4686637 0.4874052 0.8571945 0.8510946
#> [34,] 0.5568471 0.7163458 0.3822167 0.2783440 0.4820822 0.7851218 0.7376354
#> [35,] 0.5488552 0.7415461 0.3783305 0.4080782 0.3676145 0.7092207 0.7368134
#> [36,] 0.7233111 0.7882978 0.5737805 0.4304676 0.6100749 0.8593981 0.8293167
#> [37,] 0.7053992 0.7961326 0.5527973 0.5272429 0.5272736 0.8512123 0.8380591
#> [38,] 0.5946581 0.7190721 0.4536411 0.3468835 0.4784208 0.8050848 0.7813088
#> [39,] 0.6380768 0.7909794 0.4819837 0.3922210 0.5072514 0.7836452 0.8135748
#> [40,] 0.7099718 0.8087080 0.5442127 0.5075926 0.5078582 0.8539601 0.8326189
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6223367 0.7556475 0.7396432
#>  [2,] 0.6895149 0.8584875 0.7948318
#>  [3,] 0.5782088 0.7577007 0.7231926
#>  [4,] 0.6572180 0.7911228 0.7619929
#>  [5,] 0.4572436 0.7481426 0.6849422
#>  [6,] 0.6557489 0.8373842 0.7707963
#>  [7,] 0.6923756 0.8410411 0.7882670
#>  [8,] 0.6457323 0.8551747 0.7783066
#>  [9,] 0.6350753 0.8844861 0.8248991
#> [10,] 0.6813982 0.7841101 0.7576529
#> [11,] 0.5844127 0.7984024 0.7410862
#> [12,] 0.5301838 0.8074317 0.7193043
#> [13,] 0.6430401 0.8138774 0.7652336
#> [14,] 0.5171657 0.7481328 0.6757370
#> [15,] 0.6464928 0.8051483 0.7539358
#> [16,] 0.5238419 0.5853336 0.5757141
#> [17,] 0.7205411 0.8550406 0.8094092
#> [18,] 0.6098258 0.7599655 0.7000195
#> [19,] 0.6351423 0.7713160 0.7492080
#> [20,] 0.5825775 0.8193387 0.7688580
#> [21,] 0.5948460 0.7530263 0.7241587
#> [22,] 0.6057325 0.8066513 0.7437408
#> [23,] 0.7290289 0.8707136 0.8154260
#> [24,] 0.5864317 0.8530704 0.7665380
#> [25,] 0.5169768 0.5820719 0.5846587
#> [26,] 0.4572551 0.7035453 0.6358522
#> [27,] 0.7174573 0.8481214 0.8169610
#> [28,] 0.6643186 0.8346193 0.7805443
#> [29,] 0.6330584 0.7664465 0.7497457
#> [30,] 0.6880585 0.8510834 0.7907080
#> [31,] 0.6903077 0.8756165 0.8117832
#> [32,] 0.5697616 0.7618230 0.7414158
#> [33,] 0.6948606 0.8888242 0.8351699
#> [34,] 0.5522836 0.6747370 0.6483704
#> [35,] 0.5437339 0.7552917 0.6821399
#> [36,] 0.7196932 0.8307320 0.8096273
#> [37,] 0.7175265 0.8765420 0.8233389
#> [38,] 0.6115007 0.8025418 0.7529120
#> [39,] 0.6087380 0.8035625 0.7422256
#> [40,] 0.7232405 0.8654585 0.8175471
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4052747 0.5947253
#>  [2,] 0.3996545 0.6003455
#>  [3,] 0.4097023 0.5902977
#>  [4,] 0.4029938 0.5970062
#>  [5,] 0.4164355 0.5835645
#>  [6,] 0.4025381 0.5974619
#>  [7,] 0.4000476 0.5999524
#>  [8,] 0.4014934 0.5985066
#>  [9,] 0.3994436 0.6005564
#> [10,] 0.4024221 0.5975779
#> [11,] 0.4073615 0.5926385
#> [12,] 0.4089774 0.5910226
#> [13,] 0.4039640 0.5960360
#> [14,] 0.4145185 0.5854815
#> [15,] 0.4046078 0.5953922
#> [16,] 0.4203626 0.5796374
#> [17,] 0.3979136 0.6020864
#> [18,] 0.4094583 0.5905417
#> [19,] 0.4043093 0.5956907
#> [20,] 0.4030902 0.5969098
#> [21,] 0.4081473 0.5918527
#> [22,] 0.4063372 0.5936628
#> [23,] 0.3972649 0.6027351
#> [24,] 0.4053053 0.5946947
#> [25,] 0.4197728 0.5802272
#> [26,] 0.4174809 0.5825191
#> [27,] 0.3978989 0.6021011
#> [28,] 0.4016625 0.5983375
#> [29,] 0.4051193 0.5948807
#> [30,] 0.4001970 0.5998030
#> [31,] 0.3975636 0.6024364
#> [32,] 0.4063261 0.5936739
#> [33,] 0.3967582 0.6032418
#> [34,] 0.4138699 0.5861301
#> [35,] 0.4134485 0.5865515
#> [36,] 0.3977038 0.6022962
#> [37,] 0.3968558 0.6031442
#> [38,] 0.4054021 0.5945979
#> [39,] 0.4045069 0.5954931
#> [40,] 0.3973489 0.6026511
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]      [,2]          [,3]        [,4]        [,5]      [,6]
#>  [1,]  0.55942165 1.2384051 -0.1414840139 -0.75861929  0.13174437 1.5717938
#>  [2,]  0.75753031 1.2926575  0.0687865981  0.11949305 -0.11952756 1.6032746
#>  [3,]  0.20947153 0.7636865 -0.3713670692 -0.85378256 -0.20937447 1.2975127
#>  [4,]  0.67237067 1.2449092 -0.0285987618 -0.60116348  0.18722938 1.5972295
#>  [5,] -0.05599979 0.8376413 -0.6858785664 -0.96184609 -0.61511955 0.7237539
#>  [6,]  0.55893196 1.2414851  0.0001432348 -0.16408148 -0.18490045 1.3929340
#>  [7,]  0.78087807 1.0912418  0.3195541874  0.24950350  0.06783515 1.4584865
#>  [8,]  0.60036075 1.4282191 -0.1467700164 -0.07404009 -0.19362481 1.5059495
#>  [9,]  0.45263727 1.3704733 -0.2023257456 -0.30812441 -0.26747342 1.6474634
#> [10,]  0.66132246 1.4515833  0.0039177686 -0.44860043  0.20783201 1.6888221
#> [11,]  0.29174708 0.9000963 -0.3375281466 -0.74021772 -0.26475483 1.3146743
#> [12,]  0.21817415 1.1192054 -0.4067736400 -0.31834394 -0.60790066 1.0565514
#> [13,]  0.60811296 1.2961931 -0.2700561528 -0.52081344 -0.28491722 1.5338464
#> [14,]  0.22633924 0.9649504 -0.7092452606 -0.76643194 -0.66049204 0.9583745
#> [15,]  0.54971509 1.2857437 -0.2051059491 -0.45609860 -0.32535421 1.4513665
#> [16,]  0.06815335 0.6898678 -0.5477421118 -1.12804665 -0.16123683 1.0689798
#> [17,]  0.84311997 1.3999019  0.1908702888 -0.02160895  0.03723920 1.7485121
#> [18,]  0.35501737 1.1079301 -0.1656629530 -0.08311208 -0.20046178 1.1029693
#> [19,]  0.62185889 1.1951091 -0.0148331605 -0.61830387  0.11977712 1.5037514
#> [20,]  0.59645354 1.1899709 -0.1582903271 -0.69442732 -0.23222350 1.4465944
#> [21,]  0.41509289 0.8756727 -0.1636262107 -0.79088125 -0.01742894 1.3107310
#> [22,]  0.37523946 1.0733273 -0.2366439942 -0.36823281 -0.34873678 1.2840404
#> [23,]  0.83678775 1.3572212  0.2635555473  0.18186321  0.04789830 1.6840664
#> [24,]  0.30007511 1.1612012 -0.4741776666 -0.31607134 -0.66425266 1.3391666
#> [25,]  0.20522232 0.8081611 -0.5722440015 -1.10445116 -0.17219730 1.1216809
#> [26,] -0.07233339 0.9144572 -0.8244197442 -1.02297889 -0.57940678 0.9329535
#> [27,]  0.87701962 1.3768468  0.0814572848 -0.34670019  0.20512471 1.9119895
#> [28,]  0.64732316 1.2424355 -0.0758371987 -0.40828961 -0.13976441 1.5758232
#> [29,]  0.57409183 1.2600518 -0.2217147592 -0.76419721  0.03648624 1.5958077
#> [30,]  0.74776709 1.2729256  0.0412663307  0.06800706 -0.10084254 1.6083943
#> [31,]  0.75890639 1.5622414  0.1422584451  0.10851463 -0.02421515 1.6510161
#> [32,]  0.50863898 1.1133904 -0.2285647365 -0.91671131 -0.08091224 1.4055841
#> [33,]  0.71210802 1.4533058  0.0681088120 -0.12550975 -0.05038982 1.7921814
#> [34,]  0.22837594 0.9264073 -0.4801502140 -0.95269070 -0.07170209 1.2957677
#> [35,]  0.19604629 1.0540200 -0.4966406112 -0.37191568 -0.54246431 0.8916023
#> [36,]  0.96094547 1.3146953  0.2972922635 -0.27994360  0.44762698 1.8103001
#> [37,]  0.87314290 1.3622961  0.2119794146  0.10907957  0.10920282 1.7441415
#> [38,]  0.38325585 0.9398635 -0.1859695556 -0.63276634 -0.08637030 1.4183832
#> [39,]  0.56702655 1.3308390 -0.0720965264 -0.43798594  0.02900752 1.2870365
#> [40,]  0.89524724 1.4416369  0.1773139722  0.03037264  0.03143539 1.7660043
#>            [,7]        [,8]      [,9]     [,10]
#>  [1,] 1.3433768  0.49947807 1.1289630 1.0441150
#>  [2,] 1.4933988  0.79785227 1.8027838 1.3543005
#>  [3,] 1.0861510  0.31542464 1.1401146 0.9603537
#>  [4,] 1.4008554  0.65092090 1.3317064 1.1636365
#>  [5,] 0.8651175 -0.17144426 1.0887305 0.7765782
#>  [6,] 1.3546616  0.64440645 1.6388929 1.2128131
#>  [7,] 1.2988886  0.81124921 1.6659950 1.3145105
#>  [8,] 1.6097437  0.60033203 1.7757776 1.2558251
#>  [9,] 1.7358457  0.55405275 2.0356164 1.5498990
#> [10,] 1.5051108  0.76020500 1.2897806 1.1398544
#> [11,] 1.2079322  0.34091483 1.3763389 1.0516218
#> [12,] 1.1982100  0.12088220 1.4334078 0.9410132
#> [13,] 1.3820427  0.58858361 1.4754043 1.1815902
#> [14,] 0.9674896  0.06868996 1.0886783 0.7342492
#> [15,] 1.3396815  0.60365832 1.4187877 1.1197146
#> [16,] 0.8391205  0.09544007 0.3447075 0.3052036
#> [17,] 1.5719175  0.94714707 1.7746954 1.4461760
#> [18,] 1.1389907  0.44658005 1.1524903 0.8473908
#> [19,] 1.2982007  0.55434161 1.2157572 1.0943926
#> [20,] 1.4379741  0.33336323 1.5118738 1.2018740
#> [21,] 1.0765845  0.38403549 1.1148182 0.9651847
#> [22,] 1.2423743  0.42940868 1.4283963 1.0655029
#> [23,] 1.6040932  0.98970106 1.9072831 1.4856604
#> [24,] 1.3784460  0.34923336 1.7588883 1.1888648
#> [25,] 0.8023404  0.06793326 0.3312844 0.3419278
#> [26,] 0.9788198 -0.17139784 0.8642381 0.5574064
#> [27,] 1.6393713  0.93188359 1.7199422 1.4958919
#> [28,] 1.4745615  0.68259944 1.6187254 1.2688408
#> [29,] 1.3724136  0.54536015 1.1883538 1.0972565
#> [30,] 1.4525055  0.79105840 1.7431234 1.3291984
#> [31,] 1.7253345  0.80155829 1.9515590 1.4616390
#> [32,] 1.3070241  0.28087870 1.1627000 1.0533405
#> [33,] 1.7432122  0.82294257 2.0787867 1.6227197
#> [34,] 1.0337145  0.20990152 0.7296890 0.6118837
#> [35,] 1.0294713  0.17538397 1.1270369 0.7636234
#> [36,] 1.5807927  0.94294010 1.5908240 1.4475900
#> [37,] 1.6438574  0.93222518 1.9600834 1.5391348
#> [38,] 1.2733100  0.45362482 1.4022568 1.1142040
#> [39,] 1.4734077  0.44201072 1.4087105 1.0575683
#> [40,] 1.6043025  0.96059303 1.8613865 1.4998166
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.12513714 0.2583974
#>  [2,] -0.13735900 0.2695458
#>  [3,] -0.12607224 0.2391238
#>  [4,] -0.13110556 0.2619009
#>  [5,] -0.11122584 0.2261977
#>  [6,] -0.13314402 0.2617569
#>  [7,] -0.14119313 0.2640736
#>  [8,] -0.13006167 0.2691848
#>  [9,] -0.13933645 0.2684476
#> [10,] -0.12784205 0.2675412
#> [11,] -0.12874914 0.2461346
#> [12,] -0.11911325 0.2490810
#> [13,] -0.12780504 0.2611701
#> [14,] -0.10852735 0.2367895
#> [15,] -0.12641073 0.2598913
#> [16,] -0.09477658 0.2265085
#> [17,] -0.14066187 0.2735041
#> [18,] -0.11629624 0.2499088
#> [19,] -0.12903676 0.2585045
#> [20,] -0.13198612 0.2606194
#> [21,] -0.12722127 0.2444082
#> [22,] -0.12672799 0.2524002
#> [23,] -0.14267351 0.2742008
#> [24,] -0.12773135 0.2556764
#> [25,] -0.09395958 0.2297466
#> [26,] -0.10141710 0.2317061
#> [27,] -0.14148556 0.2727420
#> [28,] -0.13438624 0.2641566
#> [29,] -0.12530477 0.2588746
#> [30,] -0.13657632 0.2680680
#> [31,] -0.13847606 0.2771512
#> [32,] -0.12544435 0.2537298
#> [33,] -0.14402842 0.2749627
#> [34,] -0.10796668 0.2400233
#> [35,] -0.10906064 0.2406668
#> [36,] -0.14282759 0.2722143
#> [37,] -0.14371971 0.2748634
#> [38,] -0.13233190 0.2506741
#> [39,] -0.12466080 0.2620601
#> [40,] -0.14132664 0.2751970
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.4052747  0.4052747
#>  [2,]  0.6003455 -0.6003455
#>  [3,] -0.4097023  0.4097023
#>  [4,]  0.5970062 -0.5970062
#>  [5,] -0.4164355  0.4164355
#>  [6,]  0.5974619 -0.5974619
#>  [7,] -0.4000476  0.4000476
#>  [8,] -0.4014934  0.4014934
#>  [9,] -0.3994436  0.3994436
#> [10,] -0.4024221  0.4024221
#> [11,] -0.4073615  0.4073615
#> [12,] -0.4089774  0.4089774
#> [13,]  0.5960360 -0.5960360
#> [14,]  0.5854815 -0.5854815
#> [15,]  0.5953922 -0.5953922
#> [16,] -0.4203626  0.4203626
#> [17,]  0.6020864 -0.6020864
#> [18,] -0.4094583  0.4094583
#> [19,]  0.5956907 -0.5956907
#> [20,] -0.4030902  0.4030902
#> [21,]  0.5918527 -0.5918527
#> [22,] -0.4063372  0.4063372
#> [23,]  0.6027351 -0.6027351
#> [24,]  0.5946947 -0.5946947
#> [25,]  0.5802272 -0.5802272
#> [26,] -0.4174809  0.4174809
#> [27,]  0.6021011 -0.6021011
#> [28,]  0.5983375 -0.5983375
#> [29,] -0.4051193  0.4051193
#> [30,]  0.5998030 -0.5998030
#> [31,] -0.3975636  0.3975636
#> [32,] -0.4063261  0.4063261
#> [33,]  0.6032418 -0.6032418
#> [34,] -0.4138699  0.4138699
#> [35,]  0.5865515 -0.5865515
#> [36,] -0.3977038  0.3977038
#> [37,] -0.3968558  0.3968558
#> [38,] -0.4054021  0.4054021
#> [39,] -0.4045069  0.4045069
#> [40,]  0.6026511 -0.6026511
#> 
#> $L
#> [1] 0.6947698 0.6908602 0.7103803 0.6898227 0.7230202 0.6947867
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.6086957