Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11         V12          V13
#>  [1,] -0.3010880 -0.160856684 -0.08992049 -0.15901406 -0.157227573
#>  [2,] -0.3226879 -0.184889204 -0.07178877 -0.19395597 -0.058421720
#>  [3,] -0.3690156 -0.070190912 -0.10536673 -0.07793743 -0.207078755
#>  [4,] -0.2464786 -0.151858104 -0.06104329 -0.06344413 -0.119990038
#>  [5,] -0.3391537 -0.162609360  0.01489964 -0.01042604  0.006191302
#>  [6,] -0.2835577 -0.114042306 -0.08038660 -0.01042512  0.022599436
#>  [7,] -0.3637907 -0.158645860 -0.02828029 -0.01870364 -0.078259365
#>  [8,] -0.2028257 -0.004522204 -0.17300693 -0.03885968  0.002666218
#>  [9,] -0.2808681 -0.026208711 -0.17618762 -0.10636034 -0.089343725
#> [10,] -0.3573967 -0.040187825 -0.12523022 -0.08403942 -0.024160111
#>                 V14          V15          V16          V17          V18
#>  [1,] -0.2312875086 -0.182204058 -0.193069033 -0.005069884 -0.100241414
#>  [2,] -0.2089431731 -0.221505994 -0.185937790 -0.145066608 -0.196675443
#>  [3,] -0.0424340788 -0.108235387 -0.116064858 -0.058880709 -0.036170710
#>  [4,]  0.0002280842 -0.002698173 -0.024440404 -0.062877332 -0.027583273
#>  [5,]  0.0438121458 -0.059629049 -0.043354480 -0.034796123  0.084732467
#>  [6,] -0.0203307788 -0.019643766  0.005975726  0.023409959 -0.097096364
#>  [7,] -0.0754857107 -0.082461766  0.035830495 -0.028026686  0.007750637
#>  [8,] -0.0192027686  0.012736978 -0.136383597 -0.096472565  0.013101592
#>  [9,] -0.0892335874 -0.086949523 -0.046620735 -0.103124042  0.033834711
#> [10,] -0.1572087676 -0.054971123  0.009806776  0.031251618  0.049568131
#>                V19         V2           V20         V21         V22         V23
#>  [1,] -0.135250265 -0.1698218  0.0010680245  0.01777204  0.08007901 -0.06884674
#>  [2,] -0.007140507 -0.1860381 -0.0518045102 -0.13859233 -0.11504301  0.01053513
#>  [3,] -0.037433174 -0.2624375 -0.0092015102 -0.07824250 -0.04304847  0.14434799
#>  [4,] -0.119476709 -0.3175434  0.0226390034 -0.01416006  0.04391170  0.06222130
#>  [5,]  0.022353858 -0.3243233  0.1510770629  0.17053421  0.08463120  0.11592166
#>  [6,]  0.039060381 -0.3168055  0.0003302314  0.09520601  0.05295355  0.05173124
#>  [7,] -0.044409925 -0.2736311  0.0248501482  0.08470214  0.11078210  0.05714103
#>  [8,]  0.008251328 -0.3650596 -0.0133912959  0.02637163  0.06389912  0.16468809
#>  [9,]  0.069881759 -0.2386681  0.0042850677  0.12669036  0.15572959  0.05312961
#> [10,] -0.024897663 -0.3019536 -0.0136867805 -0.05881489  0.02488922  0.10056950
#>               V24          V25         V26          V27          V28
#>  [1,]  0.02338246 -0.005978066 -0.02112333  0.037385644 -0.022560978
#>  [2,]  0.03727396 -0.041248744  0.14684874 -0.006443934  0.078885036
#>  [3,]  0.08731237  0.102963481  0.17138768  0.235417710  0.121589146
#>  [4,]  0.03411670 -0.044276695  0.01516888  0.144228085  0.154635997
#>  [5,]  0.26599436  0.200341373  0.19223664  0.298107509  0.258648851
#>  [6,]  0.16929332  0.143183068  0.12194016  0.090064917  0.233057555
#>  [7,] -0.04357006 -0.070362425 -0.09277718 -0.116293437 -0.099748416
#>  [8,]  0.15607783  0.195087099  0.05241526  0.002126405 -0.026865286
#>  [9,]  0.16431963  0.130742957  0.13577322  0.032094325 -0.006288471
#> [10,]  0.02580554  0.220693625  0.10203728  0.140123502  0.159639668
#>                V29         V3          V30         V31          V32
#>  [1,]  0.050001517 -0.2443339  0.028796433 -0.02554197 -0.090872281
#>  [2,]  0.026986730 -0.1592515  0.082385559  0.02323321 -0.027610473
#>  [3,]  0.078622604 -0.2259737  0.047211776 -0.08291948 -0.001761884
#>  [4,]  0.079534753 -0.3348025  0.022336790  0.11548760 -0.054215028
#>  [5,]  0.254970259 -0.2621679  0.126993544  0.17057518  0.162113138
#>  [6,]  0.218285850 -0.3255170  0.245940872  0.08684914  0.035736415
#>  [7,]  0.028813842 -0.3287102  0.003153131 -0.01797765 -0.069704255
#>  [8,] -0.004512584 -0.2487696 -0.012615579 -0.05912999 -0.090070571
#>  [9,] -0.015223167 -0.2874652  0.097999469  0.09675133  0.097414581
#> [10,]  0.132355252 -0.3568243  0.160001364  0.10233388  0.096872861
#>                V33          V34         V35         V36         V37
#>  [1,] -0.082734103 -0.098850715 -0.15201112 -0.17563923 -0.04060039
#>  [2,]  0.068236963 -0.038632138  0.07943384  0.05390876 -0.07028541
#>  [3,] -0.117178026 -0.177987548 -0.18454606 -0.18200678 -0.19531915
#>  [4,] -0.033288180  0.004501489 -0.02637698  0.03013499  0.04103625
#>  [5,]  0.186390340  0.070631033  0.13183978  0.06054583  0.03164425
#>  [6,] -0.006812020  0.113370255  0.06984466 -0.02842301  0.06691835
#>  [7,] -0.066761542 -0.137062283 -0.06861061 -0.06300833 -0.05658212
#>  [8,] -0.089277274 -0.139140170 -0.13967887 -0.10436694 -0.11095890
#>  [9,] -0.001641592 -0.029345442 -0.02012091  0.05207495 -0.01214532
#> [10,]  0.089180952  0.125602148  0.01699352  0.10885491 -0.03682879
#>                V38          V39         V4          V40          V41
#>  [1,] -0.002611304 -0.012311595 -0.2011914 -0.093957689 -0.063226065
#>  [2,] -0.029941144 -0.122448312 -0.3135421 -0.040460116 -0.105440534
#>  [3,] -0.129982232 -0.203570807 -0.3400128 -0.182090497 -0.071266202
#>  [4,] -0.124476949 -0.074933836 -0.1858383  0.001498522 -0.040436291
#>  [5,]  0.086625276 -0.019763266 -0.3353283 -0.008312230 -0.045631214
#>  [6,]  0.117105102  0.142210837 -0.2329086  0.033750902  0.002829857
#>  [7,] -0.132737446 -0.078728073 -0.2324916 -0.228933974 -0.066695807
#>  [8,] -0.033728911 -0.119184350 -0.2030107 -0.136411370 -0.099969906
#>  [9,] -0.055440934  0.001578991 -0.3507976  0.033224301 -0.011501657
#> [10,] -0.040106846  0.155247044 -0.2275941  0.132693482 -0.093450554
#>                V42          V43         V44         V45         V46         V47
#>  [1,] -0.085686721 -0.169603071 -0.16894584 -0.06431776 -0.11614019 -0.12717786
#>  [2,] -0.044948739 -0.159607164 -0.04480814 -0.06288747 -0.12331726 -0.17990646
#>  [3,] -0.044685690 -0.099397689 -0.13619618 -0.08596857 -0.12339617 -0.25804920
#>  [4,] -0.048263328 -0.149429870 -0.02727810  0.01413402 -0.17433887 -0.21971058
#>  [5,] -0.061824256  0.007020328 -0.08475395 -0.15172872 -0.14409007 -0.07798169
#>  [6,] -0.005500417 -0.003924305 -0.03472997  0.01341567 -0.11704167 -0.09557666
#>  [7,] -0.261102750 -0.063070998 -0.18699263 -0.17867991 -0.19245527 -0.20463430
#>  [8,] -0.128807932 -0.101585483 -0.07978297 -0.12018490 -0.14865805 -0.08126063
#>  [9,]  0.003263966 -0.056407201 -0.09496025 -0.12857381 -0.08820441 -0.23724367
#> [10,] -0.104829287  0.027597510 -0.01178425 -0.14622920 -0.20710485 -0.22351090
#>              V48        V49         V5        V50        V51        V52
#>  [1,] -0.2682269 -0.3375087 -0.3227116 -0.3472364 -0.3562591 -0.3180750
#>  [2,] -0.2406467 -0.2365068 -0.1440821 -0.2269953 -0.2064457 -0.2330909
#>  [3,] -0.2472746 -0.3281125 -0.2399082 -0.4034755 -0.3192850 -0.3947443
#>  [4,] -0.1355282 -0.2706723 -0.1866050 -0.3405435 -0.3031997 -0.3760462
#>  [5,] -0.2718832 -0.2985028 -0.2352722 -0.3160187 -0.3612370 -0.4552803
#>  [6,] -0.1732604 -0.1955484 -0.3114060 -0.3599835 -0.2676782 -0.3985611
#>  [7,] -0.1304834 -0.1612915 -0.2004301 -0.2534136 -0.3755124 -0.2905770
#>  [8,] -0.1204383 -0.2911068 -0.2192200 -0.3390187 -0.2431613 -0.3368333
#>  [9,] -0.2507531 -0.2083332 -0.1728286 -0.3704981 -0.4518289 -0.3567898
#> [10,] -0.1896313 -0.1907845 -0.2451043 -0.2178888 -0.2995198 -0.2341970
#>              V53        V54        V55        V56        V57        V58
#>  [1,] -0.2966491 -0.2176426 -0.3202763 -0.2847969 -0.2755108 -0.2242711
#>  [2,] -0.3148112 -0.2507366 -0.3072223 -0.3688186 -0.2047628 -0.3161190
#>  [3,] -0.4087984 -0.3480908 -0.3653954 -0.2943631 -0.4136124 -0.4463563
#>  [4,] -0.3798609 -0.3910742 -0.3693275 -0.2901647 -0.4053282 -0.3862236
#>  [5,] -0.4703066 -0.4330004 -0.4343134 -0.3268742 -0.3543671 -0.4822362
#>  [6,] -0.3975243 -0.3366260 -0.3390214 -0.3383027 -0.3369846 -0.2679043
#>  [7,] -0.2370439 -0.2181796 -0.2642222 -0.2991015 -0.3639377 -0.2804707
#>  [8,] -0.2546716 -0.2470481 -0.2420230 -0.3730825 -0.3805992 -0.2876181
#>  [9,] -0.3912611 -0.3975769 -0.3040840 -0.3754735 -0.3027589 -0.3889926
#> [10,] -0.3573454 -0.3776133 -0.3992722 -0.4170278 -0.3944290 -0.2632620
#>              V59         V6        V60          V7          V8          V9
#>  [1,] -0.1783826 -0.2634168 -0.2407549 -0.09047762 -0.07966519 -0.13086070
#>  [2,] -0.2452626 -0.1036475 -0.2522848 -0.22076886 -0.07674021 -0.19892636
#>  [3,] -0.3316111 -0.1566346 -0.4103213 -0.11946523 -0.19969876 -0.09078401
#>  [4,] -0.4241069 -0.2046936 -0.3902787 -0.18329310 -0.20420512 -0.16682821
#>  [5,] -0.4605466 -0.2066159 -0.3679400 -0.14698994 -0.13674235 -0.16606252
#>  [6,] -0.3450567 -0.2152956 -0.3643289 -0.22160418 -0.19019889 -0.14397318
#>  [7,] -0.3049291 -0.1866106 -0.2821344 -0.22257957 -0.24787657 -0.05419275
#>  [8,] -0.2959087 -0.2285378 -0.2928695 -0.20835058 -0.21517405 -0.02432065
#>  [9,] -0.4241405 -0.2804579 -0.3883818 -0.14197650 -0.24464361 -0.16527860
#> [10,] -0.4121604 -0.2560416 -0.3812614 -0.24942040 -0.24001206 -0.15934703
#> 
#> $W[[2]]
#>            [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.02604379 -0.06985737 -0.04219513  0.05376337 0.031351067  0.058192057
#> [2,] 0.02709650 -0.02617725 -0.02516302 -0.01507150 0.002615008 -0.002188266
#>             [,7]       [,8]        [,9]       [,10]
#> [1,] -0.02891906 0.07219838 -0.01208698 -0.01647457
#> [2,]  0.03781036 0.03479618  0.01663213  0.03291725
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 2.978500e-05 0.0003514438 0.0003290216 0.0003144261 0.0003831013
#>  [2,] 6.625152e-05 0.0006167585 0.0006663827 0.0007018463 0.0007136037
#>  [3,] 1.128474e-04 0.0010522394 0.0011585992 0.0012388757 0.0012157141
#>  [4,] 4.143468e-05 0.0009744235 0.0007834831 0.0006819664 0.0008765927
#>  [5,] 5.109943e-05 0.0006769015 0.0006616695 0.0006559841 0.0006669609
#>  [6,] 6.801193e-05 0.0010883080 0.0010157359 0.0009754860 0.0010395836
#>  [7,] 6.928819e-05 0.0005284483 0.0006479194 0.0007212917 0.0006691181
#>  [8,] 6.430354e-05 0.0009017118 0.0008263681 0.0007765643 0.0009018375
#>  [9,] 1.246210e-04 0.0012713697 0.0013274426 0.0013768070 0.0013743692
#> [10,] 1.352918e-04 0.0012286711 0.0013316293 0.0014007814 0.0013724883
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0004955722 0.0005673742 0.0007408730 0.0009282894 0.0010161402
#>  [2,] 0.0007291153 0.0007492802 0.0008941361 0.0010014058 0.0011498610
#>  [3,] 0.0011782386 0.0012017871 0.0014801913 0.0016647047 0.0018831971
#>  [4,] 0.0012944077 0.0016173212 0.0023438624 0.0031091596 0.0032984488
#>  [5,] 0.0007281221 0.0008101559 0.0011073457 0.0013219967 0.0014124238
#>  [6,] 0.0012185032 0.0014029981 0.0019752549 0.0024758046 0.0026388776
#>  [7,] 0.0005739369 0.0005406545 0.0005647124 0.0005170838 0.0006651856
#>  [8,] 0.0011620749 0.0013633791 0.0018596192 0.0023669739 0.0025453157
#>  [9,] 0.0014180294 0.0014963583 0.0019106538 0.0022075209 0.0024346668
#> [10,] 0.0013692272 0.0014261104 0.0017795758 0.0019833585 0.0022018389
#>                V19           V2         V20         V21         V22         V23
#>  [1,] 0.0009889606 4.946944e-05 0.001020012 0.001227113 0.001315345 0.001298611
#>  [2,] 0.0013252028 9.716067e-05 0.001444746 0.001639460 0.001722454 0.001749789
#>  [3,] 0.0022304295 1.639629e-04 0.002434351 0.002700285 0.002808351 0.002782774
#>  [4,] 0.0029086890 9.871053e-05 0.002776630 0.003400473 0.003649182 0.003446353
#>  [5,] 0.0014574980 8.574140e-05 0.001507324 0.001751753 0.001760012 0.001579536
#>  [6,] 0.0025766283 1.234859e-04 0.002614591 0.003058587 0.003148310 0.002886094
#>  [7,] 0.0009996687 9.390919e-05 0.001151660 0.001218973 0.001225270 0.001256382
#>  [8,] 0.0024201511 1.131005e-04 0.002433366 0.002901330 0.003103087 0.002997097
#>  [9,] 0.0027093633 1.883597e-04 0.002912674 0.003283012 0.003385420 0.003278233
#> [10,] 0.0025898232 1.969777e-04 0.002857838 0.003162392 0.003200901 0.003072825
#>               V24         V25         V26         V27         V28         V29
#>  [1,] 0.001356335 0.001414964 0.001404516 0.001378243 0.001342148 0.001245632
#>  [2,] 0.001859916 0.001918084 0.001971820 0.002028147 0.001974796 0.001781674
#>  [3,] 0.002865233 0.002877422 0.002985971 0.003150401 0.003120731 0.002825432
#>  [4,] 0.003555452 0.003631525 0.003413804 0.003289291 0.003325977 0.003201521
#>  [5,] 0.001602949 0.001630162 0.001622885 0.001653355 0.001676334 0.001519043
#>  [6,] 0.002919056 0.002957845 0.002890086 0.002908360 0.002950464 0.002702933
#>  [7,] 0.001320408 0.001307233 0.001384290 0.001478532 0.001417518 0.001231760
#>  [8,] 0.003084499 0.003149659 0.003068659 0.003037407 0.003026725 0.002840789
#>  [9,] 0.003377392 0.003419155 0.003502293 0.003652366 0.003666992 0.003354671
#> [10,] 0.003124717 0.003124236 0.003231733 0.003395324 0.003407582 0.003089753
#>                 V3         V30          V31          V32          V33
#>  [1,] 6.045996e-05 0.001138851 0.0009136882 0.0006474476 0.0007232959
#>  [2,] 1.014310e-04 0.001584576 0.0012741539 0.0010639032 0.0009713154
#>  [3,] 1.697523e-04 0.002528387 0.0020070206 0.0017194723 0.0015051481
#>  [4,] 1.616798e-04 0.003203194 0.0025775592 0.0015188269 0.0019785020
#>  [5,] 1.072701e-04 0.001512389 0.0011950994 0.0008316781 0.0008506589
#>  [6,] 1.690836e-04 0.002661294 0.0020980377 0.0013820277 0.0015426493
#>  [7,] 8.621197e-05 0.001025688 0.0008012482 0.0008102133 0.0005855697
#>  [8,] 1.496515e-04 0.002713435 0.0021677501 0.0014646833 0.0016721248
#>  [9,] 2.040920e-04 0.003112476 0.0024808615 0.0019947755 0.0018571594
#> [10,] 2.028651e-04 0.002831274 0.0022262425 0.0018829895 0.0016796397
#>                V34          V35           V36           V37          V38
#>  [1,] 0.0008602290 9.931871e-04  0.0009909250  9.534234e-04 0.0007164211
#>  [2,] 0.0008676393 8.045792e-04  0.0007393875  7.864598e-04 0.0007370912
#>  [3,] 0.0013370149 1.197920e-03  0.0010389266  1.133293e-03 0.0011561834
#>  [4,] 0.0027929819 3.542234e-03  0.0036860646  3.555252e-03 0.0024840601
#>  [5,] 0.0010034176 1.157085e-03  0.0011244345  1.145533e-03 0.0009593404
#>  [6,] 0.0019497623 2.327089e-03  0.0023289388  2.331361e-03 0.0018008020
#>  [7,] 0.0002700990 2.750249e-06 -0.0001256244 -5.651096e-06 0.0002454746
#>  [8,] 0.0020791796 2.473939e-03  0.0024984427  2.434513e-03 0.0018056128
#>  [9,] 0.0018625188 1.876911e-03  0.0017254384  1.793654e-03 0.0016365532
#> [10,] 0.0015871409 1.480351e-03  0.0012542348  1.334813e-03 0.0013687000
#>                V39           V4          V40          V41          V42
#>  [1,] 0.0005006518 8.366612e-05 0.0006184743 0.0005622866 0.0004257532
#>  [2,] 0.0006641988 1.331397e-04 0.0007036077 0.0007289896 0.0007491524
#>  [3,] 0.0010931551 2.308598e-04 0.0011077785 0.0011874332 0.0012213159
#>  [4,] 0.0015540422 2.304711e-04 0.0021212992 0.0017730462 0.0009588043
#>  [5,] 0.0007487815 1.541233e-04 0.0008782908 0.0008470380 0.0006323952
#>  [6,] 0.0012789425 2.482179e-04 0.0016026743 0.0014692464 0.0009683866
#>  [7,] 0.0004134115 1.094692e-04 0.0003192898 0.0004224678 0.0006301672
#>  [8,] 0.0012250384 2.123326e-04 0.0015577290 0.0013982495 0.0009554798
#>  [9,] 0.0013995988 2.840998e-04 0.0014998813 0.0015199896 0.0013909711
#> [10,] 0.0012712831 2.801751e-04 0.0012720223 0.0013645252 0.0013611528
#>                V43          V44          V45          V46          V47
#>  [1,] 0.0003566590 0.0003124152 0.0002034291 1.395623e-04 0.0001340447
#>  [2,] 0.0007094440 0.0006608691 0.0005843909 4.738218e-04 0.0003581866
#>  [3,] 0.0012008072 0.0011343499 0.0010408603 8.486203e-04 0.0006173059
#>  [4,] 0.0007389672 0.0006656977 0.0002300732 2.010807e-05 0.0001524491
#>  [5,] 0.0005895328 0.0005847592 0.0004563713 3.137397e-04 0.0002395485
#>  [6,] 0.0008555907 0.0008349202 0.0005760835 3.617566e-04 0.0003162365
#>  [7,] 0.0006559871 0.0006165620 0.0006176376 5.406554e-04 0.0003816112
#>  [8,] 0.0008098770 0.0007306415 0.0004566833 2.871691e-04 0.0002885390
#>  [9,] 0.0013418943 0.0012783033 0.0011186510 8.726847e-04 0.0006414510
#> [10,] 0.0013335689 0.0012639962 0.0011676717 9.499282e-04 0.0006713994
#>                V48          V49           V5          V50           V51
#>  [1,] 0.0001170520 6.874168e-05 0.0001273699 3.293172e-05  1.281561e-05
#>  [2,] 0.0002702793 1.583325e-04 0.0001924638 5.256662e-05  4.160991e-05
#>  [3,] 0.0004501674 2.693796e-04 0.0003274814 8.986598e-05  7.483583e-05
#>  [4,] 0.0001878360 1.200347e-04 0.0003832150 8.384634e-05 -4.602400e-06
#>  [5,] 0.0001925149 1.209093e-04 0.0002271289 4.977159e-05  2.361583e-05
#>  [6,] 0.0002712917 1.749835e-04 0.0003793347 8.250269e-05  2.678385e-05
#>  [7,] 0.0002640764 1.571906e-04 0.0001481314 4.098194e-05  4.984444e-05
#>  [8,] 0.0002574428 1.569757e-04 0.0003256599 7.879307e-05  2.383812e-05
#>  [9,] 0.0004850935 2.896220e-04 0.0004058686 1.057134e-04  7.405180e-05
#> [10,] 0.0004886598 2.924271e-04 0.0003872069 1.033713e-04  8.338670e-05
#>                 V52          V53          V54          V55          V56
#>  [1,]  1.013475e-05 1.296822e-05 1.352255e-05 1.195455e-05 1.502215e-05
#>  [2,]  3.475129e-05 2.284956e-05 2.366326e-05 1.724647e-05 2.130383e-05
#>  [3,]  6.136904e-05 3.975023e-05 3.900865e-05 2.779195e-05 3.451469e-05
#>  [4,] -6.375398e-06 2.677765e-05 2.362242e-05 2.795522e-05 4.103034e-05
#>  [5,]  1.888560e-05 2.106638e-05 1.805221e-05 1.639024e-05 2.282494e-05
#>  [6,]  2.005454e-05 3.163327e-05 2.697113e-05 2.641013e-05 3.670463e-05
#>  [7,]  4.182847e-05 2.092713e-05 2.177639e-05 1.302325e-05 1.553852e-05
#>  [8,]  1.806219e-05 2.889981e-05 2.786550e-05 2.627543e-05 3.501142e-05
#>  [9,]  6.186071e-05 4.580313e-05 4.333644e-05 3.396732e-05 4.325769e-05
#> [10,]  7.019247e-05 4.773786e-05 4.533295e-05 3.408062e-05 4.192349e-05
#>                V57          V58          V59           V6          V60
#>  [1,] 1.789790e-05 1.272064e-05 1.129000e-05 0.0001789737 1.116593e-05
#>  [2,] 1.727973e-05 1.670290e-05 1.512468e-05 0.0002550496 1.453737e-05
#>  [3,] 2.601333e-05 2.840859e-05 2.603214e-05 0.0004238517 2.437978e-05
#>  [4,] 6.086730e-05 3.745216e-05 3.419816e-05 0.0005569308 3.374179e-05
#>  [5,] 2.322870e-05 1.944930e-05 1.989861e-05 0.0002828551 1.850160e-05
#>  [6,] 4.445285e-05 3.242459e-05 3.140506e-05 0.0004999168 3.047084e-05
#>  [7,] 5.055132e-06 1.174103e-05 1.068457e-05 0.0001796233 9.486237e-06
#>  [8,] 4.487880e-05 3.099313e-05 2.777472e-05 0.0004567566 2.762154e-05
#>  [9,] 3.797611e-05 3.594640e-05 3.352699e-05 0.0005270697 3.179214e-05
#> [10,] 3.250030e-05 3.457260e-05 3.259658e-05 0.0004884029 3.073439e-05
#>                 V7           V8           V9
#>  [1,] 0.0002528217 0.0002696710 0.0003246098
#>  [2,] 0.0003146458 0.0003519022 0.0005394165
#>  [3,] 0.0004945006 0.0005672233 0.0009106404
#>  [4,] 0.0008179881 0.0008756553 0.0009326514
#>  [5,] 0.0003628333 0.0004206297 0.0006034901
#>  [6,] 0.0006623678 0.0007453592 0.0009847619
#>  [7,] 0.0001851758 0.0002196139 0.0004327111
#>  [8,] 0.0006352482 0.0006928820 0.0008335863
#>  [9,] 0.0006425644 0.0007440028 0.0011238175
#> [10,] 0.0005689165 0.0006725148 0.0010773377
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.01412527  0.02096404  0.01057158  0.02689761  0.03436146  0.03137878
#> [2,] -0.01412527 -0.02096404 -0.01057158 -0.02689761 -0.03436146 -0.03137878
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01638913  0.01656986  0.02995042  0.03254792
#> [2,] -0.01638913 -0.01656986 -0.02995042 -0.03254792
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.06381982 0.22482034 0.45392179 0.46865886 0.93556719 0.59384828
#>  [7] 0.14387791 0.35446243 0.57649001 0.62145271
#> 
#> $B[[2]]
#> [1]  0.1261433 -0.0148104
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.001937315 0.002532605 0.003925209 0.005320652 0.002452894 0.004372775
#>  [7] 0.001645135 0.004456324 0.004833018 0.004436561
#> 
#> $vB[[2]]
#> [1]  0.03685533 -0.03685533
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#>  [2,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#>  [3,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#>  [4,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#>  [5,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#>  [6,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#>  [7,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#>  [8,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [9,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [10,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [11,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [12,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [13,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [14,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [15,] 0.0130 0.1483 0.1532 0.1100 0.0890 0.1236 0.1197 0.1145 0.2137 0.2838
#> [16,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [17,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [18,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [19,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [20,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [21,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [22,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [23,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [24,] 0.0209 0.1853 0.1748 0.1556 0.1476 0.1378 0.2584 0.3827 0.4784 0.5360
#> [25,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [26,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [27,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [28,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [29,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [30,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [31,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [32,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [33,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [34,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#> [35,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [36,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [37,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [38,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [39,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [40,] 0.0201 0.0279 0.2251 0.2615 0.1770 0.3709 0.4533 0.5553 0.4616 0.3797
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#>  [2,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#>  [3,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#>  [4,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#>  [5,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#>  [6,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#>  [7,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#>  [8,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [9,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [10,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [11,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [12,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [13,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [14,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [15,] 0.3640 0.0006 0.5430 0.6673 0.7979 0.9273 0.9027 0.9192 1.0000 0.9821
#> [16,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [17,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [18,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [19,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [20,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [21,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [22,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [23,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [24,] 0.6192 0.0278 0.7912 0.9264 1.0000 0.9080 0.7435 0.5557 0.3172 0.1295
#> [25,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [26,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [27,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [28,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [29,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [30,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [31,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [32,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [33,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [34,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#> [35,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [36,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [37,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [38,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [39,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [40,] 0.3450 0.0423 0.2665 0.2395 0.1127 0.2556 0.5169 0.3779 0.4082 0.5353
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#>  [2,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#>  [3,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#>  [4,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#>  [5,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#>  [6,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#>  [7,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#>  [8,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [9,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [10,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [11,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [12,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [13,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [14,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [15,] 0.9092 0.8184 0.0088 0.6962 0.5900 0.5447 0.5142 0.5389 0.5531 0.5318
#> [16,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [17,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [18,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [19,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [20,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [21,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [22,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [23,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [24,] 0.0598 0.2722 0.0115 0.3616 0.3293 0.4855 0.3936 0.1845 0.0342 0.2489
#> [25,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [26,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [27,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [28,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [29,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [30,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [31,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [32,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [33,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [34,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#> [35,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [36,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [37,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [38,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [39,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [40,] 0.5116 0.4544 0.0554 0.4258 0.3869 0.3939 0.4661 0.3974 0.2194 0.1816
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#>  [2,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#>  [3,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#>  [4,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#>  [5,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#>  [6,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#>  [7,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#>  [8,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [9,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [10,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [11,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [12,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [13,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [14,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [15,] 0.4826 0.3790 0.1831 0.0456 0.1750 0.1679 0.0674 0.0609 0.0375 0.0533
#> [16,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [17,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [18,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [19,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [20,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [21,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [22,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [23,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [24,] 0.3837 0.3514 0.2654 0.0445 0.1760 0.1599 0.0866 0.0590 0.0813 0.0492
#> [25,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [26,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [27,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [28,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [29,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [30,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [31,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [32,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [33,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [34,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#> [35,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [36,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [37,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [38,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [39,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [40,] 0.1023 0.2108 0.3253 0.0783 0.3697 0.2912 0.3010 0.2563 0.1927 0.2062
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#>  [2,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#>  [3,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#>  [4,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#>  [5,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#>  [6,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#>  [7,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#>  [8,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [9,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [10,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [11,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [12,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [13,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [14,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [15,] 0.0278 0.0179 0.0114 0.0073 0.0525 0.0116 0.0092 0.0078 0.0041 0.0013
#> [16,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [17,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [18,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [19,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [20,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [21,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [22,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [23,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [24,] 0.0417 0.0495 0.0367 0.0115 0.0427 0.0118 0.0133 0.0096 0.0014 0.0049
#> [25,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [26,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [27,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [28,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [29,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [30,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [31,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [32,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [33,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [34,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#> [35,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [36,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [37,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [38,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [39,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [40,] 0.1751 0.0841 0.1035 0.0641 0.0620 0.0153 0.0081 0.0191 0.0182 0.0160
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#>  [2,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#>  [3,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#>  [4,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#>  [5,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#>  [6,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#>  [7,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#>  [8,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [9,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [10,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [11,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [12,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [13,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [14,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [15,] 0.0011 0.0045 0.0039 0.0022 0.0023 0.0778 0.0016 0.0931 0.0941 0.1711
#> [16,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [17,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [18,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [19,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [20,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [21,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [22,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [23,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [24,] 0.0039 0.0029 0.0078 0.0047 0.0021 0.0766 0.0011 0.1458 0.1430 0.1894
#> [25,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [26,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [27,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [28,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [29,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [30,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [31,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [32,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [33,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [34,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#> [35,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [36,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [37,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [38,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [39,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [40,] 0.0290 0.0090 0.0242 0.0224 0.0190 0.0871 0.0096 0.1201 0.2707 0.1206
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3449533 0.4506011 0.5941701 0.6223185 0.9269708 0.8386279 0.3161258
#>  [2,] 0.2947026 0.4045494 0.5388147 0.5873543 0.9238959 0.8240237 0.2903800
#>  [3,] 0.2014120 0.3280480 0.2674141 0.4835569 0.8677232 0.8121441 0.1743788
#>  [4,] 0.2434784 0.2824161 0.4833458 0.5552065 0.9203094 0.8177291 0.3298862
#>  [5,] 0.2435070 0.3015346 0.4552959 0.5422645 0.8829678 0.7598493 0.3156337
#>  [6,] 0.2568275 0.4204751 0.4084163 0.6117276 0.9496372 0.8786276 0.2441064
#>  [7,] 0.2209910 0.3575496 0.3527618 0.5321522 0.9122359 0.8074517 0.2452911
#>  [8,] 0.2964576 0.3303829 0.4712414 0.5684201 0.8931883 0.7775586 0.4025815
#>  [9,] 0.2639943 0.3389158 0.5236531 0.5766869 0.9261290 0.8211513 0.2827927
#> [10,] 0.2962611 0.3662817 0.5941784 0.6104870 0.9426738 0.8459728 0.3318829
#> [11,] 0.3388535 0.4963687 0.5565385 0.6692234 0.9609347 0.8867560 0.3296085
#> [12,] 0.1717515 0.2922624 0.2972972 0.4406436 0.8320143 0.7402387 0.1747267
#> [13,] 0.3384667 0.3603524 0.5379425 0.5857001 0.8852239 0.7870196 0.3869652
#> [14,] 0.3028654 0.4810892 0.4830278 0.6253257 0.9412454 0.8535494 0.3131970
#> [15,] 0.3579489 0.4903112 0.6048144 0.6677211 0.9560948 0.8730147 0.3523584
#> [16,] 0.3498683 0.4712873 0.6440129 0.6438770 0.9428113 0.8509636 0.3449012
#> [17,] 0.2166478 0.3019205 0.2991822 0.4902558 0.9011256 0.7833558 0.2526069
#> [18,] 0.2752259 0.4641799 0.3320441 0.5912249 0.8827164 0.8112018 0.2656195
#> [19,] 0.3212331 0.4164518 0.5665060 0.6280205 0.9465260 0.8644034 0.3242084
#> [20,] 0.2118967 0.2119756 0.3936439 0.4595623 0.8423165 0.6841732 0.3134221
#> [21,] 0.2177670 0.3475587 0.3767237 0.5330351 0.9083789 0.8180877 0.2099576
#> [22,] 0.2661582 0.3138099 0.5475769 0.5629000 0.9163840 0.7959403 0.3127225
#> [23,] 0.2283020 0.3017668 0.4441170 0.5532959 0.9268628 0.8385260 0.2272057
#> [24,] 0.3494708 0.3524857 0.4869547 0.5582226 0.8828613 0.7616395 0.4321271
#> [25,] 0.2272973 0.2686186 0.4052630 0.5236491 0.8960793 0.7838197 0.2949595
#> [26,] 0.2977062 0.3355422 0.4395517 0.5444746 0.8947648 0.7788709 0.3711067
#> [27,] 0.2306175 0.2604265 0.3561864 0.5016240 0.8668624 0.7336912 0.3833746
#> [28,] 0.2583740 0.3086087 0.4791614 0.5468345 0.9173087 0.8204316 0.2787233
#> [29,] 0.2778885 0.3001377 0.4920436 0.5954026 0.9199041 0.8203792 0.3804346
#> [30,] 0.3755331 0.4489777 0.6416744 0.6245729 0.9300757 0.8308403 0.3545773
#> [31,] 0.3396394 0.5149168 0.5856271 0.6624043 0.9475266 0.8675118 0.3240138
#> [32,] 0.3152807 0.4333809 0.5670940 0.6032412 0.9188772 0.8245840 0.2972129
#> [33,] 0.2170845 0.2954124 0.3996701 0.5221197 0.9097172 0.8181788 0.2217846
#> [34,] 0.2597527 0.3819996 0.4083228 0.5580199 0.9147485 0.8202312 0.2505744
#> [35,] 0.1719530 0.2026768 0.2944806 0.4621966 0.8830109 0.7784079 0.2354004
#> [36,] 0.3339081 0.3815129 0.6387802 0.6440431 0.9440023 0.8425772 0.4004621
#> [37,] 0.3724010 0.4356944 0.5687682 0.6370381 0.9276567 0.8362565 0.3688519
#> [38,] 0.2301272 0.2939960 0.4436614 0.5609903 0.8971627 0.7919747 0.2856197
#> [39,] 0.2463174 0.3450100 0.4411014 0.5047906 0.8726813 0.7556255 0.2573818
#> [40,] 0.2744794 0.3605588 0.4450982 0.5376776 0.8460880 0.7414673 0.2957170
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5258246 0.7016620 0.7897361
#>  [2,] 0.5160108 0.7193884 0.7845025
#>  [3,] 0.3095382 0.6292537 0.7234010
#>  [4,] 0.5000063 0.6750282 0.7435748
#>  [5,] 0.4901325 0.6662832 0.7102703
#>  [6,] 0.4055471 0.7220496 0.8467355
#>  [7,] 0.3915890 0.6353382 0.7458499
#>  [8,] 0.5082604 0.6950242 0.7237692
#>  [9,] 0.5317450 0.6901928 0.7546255
#> [10,] 0.5646969 0.7277863 0.7880497
#> [11,] 0.5270684 0.7749052 0.8559313
#> [12,] 0.2954368 0.5054437 0.6309856
#> [13,] 0.5222188 0.6498668 0.7245470
#> [14,] 0.4886913 0.7504795 0.8260405
#> [15,] 0.5712091 0.7743508 0.8381097
#> [16,] 0.5796498 0.7333851 0.8188636
#> [17,] 0.3739064 0.6541694 0.7230566
#> [18,] 0.3333314 0.6382713 0.7735871
#> [19,] 0.5653807 0.7461313 0.8091041
#> [20,] 0.4606057 0.5865253 0.6099337
#> [21,] 0.3794977 0.6296669 0.7265910
#> [22,] 0.5396230 0.6853828 0.7211589
#> [23,] 0.4362458 0.6581091 0.7599508
#> [24,] 0.5648975 0.7223030 0.7137952
#> [25,] 0.4801322 0.6701862 0.7068729
#> [26,] 0.5111456 0.7093531 0.7216821
#> [27,] 0.4753692 0.6458761 0.6799026
#> [28,] 0.5177682 0.6829577 0.7549018
#> [29,] 0.5065024 0.6796155 0.7518121
#> [30,] 0.5798532 0.7281461 0.7868564
#> [31,] 0.5325018 0.7470278 0.8407880
#> [32,] 0.4868385 0.6700926 0.7725423
#> [33,] 0.3772307 0.6381216 0.7053731
#> [34,] 0.4393468 0.7060471 0.7709657
#> [35,] 0.3543181 0.5665691 0.6669894
#> [36,] 0.6099682 0.7526688 0.7998774
#> [37,] 0.5474712 0.7307595 0.7826676
#> [38,] 0.4844240 0.6607308 0.7157428
#> [39,] 0.4630621 0.6224085 0.6834349
#> [40,] 0.4238589 0.5944396 0.6925581
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6355876 0.3644124
#>  [2,] 0.6339075 0.3660925
#>  [3,] 0.6251429 0.3748571
#>  [4,] 0.6313579 0.3686421
#>  [5,] 0.6282080 0.3717920
#>  [6,] 0.6351751 0.3648249
#>  [7,] 0.6285285 0.3714715
#>  [8,] 0.6299054 0.3700946
#>  [9,] 0.6333878 0.3666122
#> [10,] 0.6362365 0.3637635
#> [11,] 0.6396407 0.3603593
#> [12,] 0.6193756 0.3806244
#> [13,] 0.6308353 0.3691647
#> [14,] 0.6359384 0.3640616
#> [15,] 0.6397386 0.3602614
#> [16,] 0.6379643 0.3620357
#> [17,] 0.6260280 0.3739720
#> [18,] 0.6289392 0.3710608
#> [19,] 0.6377359 0.3622641
#> [20,] 0.6210689 0.3789311
#> [21,] 0.6287474 0.3712526
#> [22,] 0.6318353 0.3681647
#> [23,] 0.6316081 0.3683919
#> [24,] 0.6301237 0.3698763
#> [25,] 0.6285402 0.3714598
#> [26,] 0.6296628 0.3703372
#> [27,] 0.6247178 0.3752822
#> [28,] 0.6320026 0.3679974
#> [29,] 0.6325606 0.3674394
#> [30,] 0.6364989 0.3635011
#> [31,] 0.6385017 0.3614983
#> [32,] 0.6334754 0.3665246
#> [33,] 0.6283597 0.3716403
#> [34,] 0.6313304 0.3686696
#> [35,] 0.6233819 0.3766181
#> [36,] 0.6378166 0.3621834
#> [37,] 0.6362316 0.3637684
#> [38,] 0.6300201 0.3699799
#> [39,] 0.6263616 0.3736384
#> [40,] 0.6253446 0.3746554
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]        [,2]        [,3]         [,4]     [,5]     [,6]
#>  [1,] -0.6412976 -0.19824222  0.38123167  0.499400850 2.541063 1.648054
#>  [2,] -0.8726530 -0.38654477  0.15557193  0.353039031 2.496497 1.543850
#>  [3,] -1.3774923 -0.71702703 -1.00778214 -0.065796130 1.880976 1.464002
#>  [4,] -1.1337031 -0.93250819 -0.06664146  0.221730027 2.446558 1.501037
#>  [5,] -1.1335476 -0.84000072 -0.17929507  0.169462547 2.020839 1.151854
#>  [6,] -1.0625237 -0.32082338 -0.37051627  0.454580056 2.936828 1.979497
#>  [7,] -1.2599006 -0.58601546 -0.60692137  0.128786570 2.341246 1.433536
#>  [8,] -0.8642238 -0.70645372 -0.11516136  0.275408060 2.123730 1.251496
#>  [9,] -1.0253102 -0.66812934  0.09468299  0.309187552 2.528693 1.524167
#> [10,] -0.8651664 -0.54820175  0.38126593  0.449359678 2.799962 1.703358
#> [11,] -0.6684076 -0.01452528  0.22712542  0.704674649 3.202671 2.058025
#> [12,] -1.5732645 -0.88442123 -0.86020180 -0.238550315 1.599971 1.047209
#> [13,] -0.6701346 -0.57383476  0.15206239  0.346217577 2.042857 1.307053
#> [14,] -0.8336899 -0.07567943 -0.06791472  0.512215470 2.773834 1.762715
#> [15,] -0.5842777 -0.03876003  0.42556601  0.697895977 3.080825 1.927881
#> [16,] -0.6196184 -0.11497745  0.59282442  0.592231450 2.802509 1.742178
#> [17,] -1.2853096 -0.83816914 -0.85119530 -0.038981617 2.209795 1.285330
#> [18,] -0.9682680 -0.14352627 -0.69895438  0.369031190 2.018409 1.457838
#> [19,] -0.7481107 -0.33735652  0.26760954  0.523734112 2.873602 1.852356
#> [20,] -1.3135306 -1.31305772 -0.43202066 -0.162105037 1.675566 0.773017
#> [21,] -1.2787268 -0.62978749 -0.50347801  0.132333369 2.294000 1.503445
#> [22,] -1.0142024 -0.78236721  0.19088524  0.252939815 2.394201 1.361111
#> [23,] -1.2179241 -0.83889843 -0.22446981  0.213996647 2.539469 1.647302
#> [24,] -0.6213660 -0.60813073 -0.05219311  0.233951617 2.019809 1.161689
#> [25,] -1.2236357 -1.00164251 -0.38358307  0.094667189 2.154401 1.288066
#> [26,] -0.8582446 -0.68322377 -0.24298146  0.178369998 2.140363 1.259099
#> [27,] -1.2048275 -1.04375287 -0.59195494  0.006496141 1.873497 1.013432
#> [28,] -1.0544371 -0.80663193 -0.08340286  0.187888768 2.406330 1.519275
#> [29,] -0.9549595 -0.84664208 -0.03182811  0.386345207 2.441045 1.518919
#> [30,] -0.5085516 -0.20480223  0.58263884  0.509003922 2.587853 1.591594
#> [31,] -0.6649016  0.05968479  0.34591686  0.674026890 2.893548 1.879136
#> [32,] -0.7755457 -0.26807019  0.27000432  0.418988635 2.427189 1.547719
#> [33,] -1.2827380 -0.86924055 -0.40684005  0.088536692 2.310187 1.504057
#> [34,] -1.0472543 -0.48106971 -0.37090319  0.233129949 2.373044 1.517914
#> [35,] -1.5718487 -1.36964783 -0.87372119 -0.151502674 2.021256 1.256412
#> [36,] -0.6905620 -0.48313168  0.57007373  0.592955865 2.824817 1.677530
#> [37,] -0.5219303 -0.25865503  0.27682702  0.562531571 2.551238 1.630635
#> [38,] -1.2075929 -0.87605456 -0.22631535  0.245182308 2.166089 1.336870
#> [39,] -1.1183502 -0.64104663 -0.23669324  0.019163028 1.924877 1.128844
#> [40,] -0.9720134 -0.57293965 -0.22049605  0.150996472 1.704242 1.053609
#>             [,7]          [,8]       [,9]     [,10]
#>  [1,] -0.7716337  1.033903e-01 0.85522493 1.3233355
#>  [2,] -0.8935393  6.406511e-02 0.94143012 1.2921007
#>  [3,] -1.5549061 -8.022794e-01 0.52901670 0.9613950
#>  [4,] -0.7086996  2.524512e-05 0.73101600 1.0646325
#>  [5,] -0.7739110 -3.947509e-02 0.69142207 0.8966973
#>  [6,] -1.1302962 -3.824044e-01 0.95465138 1.7092229
#>  [7,] -1.1238865 -4.406379e-01 0.55518713 1.0765993
#>  [8,] -0.3947202  3.304470e-02 0.82371429 0.9632357
#>  [9,] -0.9306510  1.271509e-01 0.80102101 1.1234361
#> [10,] -0.6996814  2.602464e-01 0.98341990 1.3132092
#> [11,] -0.7099561  1.083794e-01 1.23621923 1.7818996
#> [12,] -1.5524914 -8.691232e-01 0.02177574 0.5364473
#> [13,] -0.4600871  8.893373e-02 0.61845370 0.9671297
#> [14,] -0.7852151 -4.524244e-02 1.10117152 1.5578215
#> [15,] -0.6086888  2.867859e-01 1.23304316 1.6442303
#> [16,] -0.6415282  3.213360e-01 1.01186554 1.5086674
#> [17,] -1.0847566 -5.154945e-01 0.63741723 0.9596745
#> [18,] -1.0169625 -6.931557e-01 0.56786874 1.2286779
#> [19,] -0.7344984  2.630287e-01 1.07808414 1.4441994
#> [20,] -0.7841690 -1.579043e-01 0.34961933 0.4470337
#> [21,] -1.3251810 -4.916808e-01 0.53078827 0.9773950
#> [22,] -0.7874221  1.588250e-01 0.77862064 0.9502175
#> [23,] -1.2241574 -2.564124e-01 0.65487902 1.1524095
#> [24,] -0.2731781  2.610626e-01 0.95591411 0.9138883
#> [25,] -0.8714174 -7.951300e-02 0.70902721 0.8802444
#> [26,] -0.5274720  4.458989e-02 0.89224441 0.9528205
#> [27,] -0.4752490 -9.860284e-02 0.60096067 0.7533244
#> [28,] -0.9508032  7.110292e-02 0.76739787 1.1249289
#> [29,] -0.4877039  2.601116e-02 0.75200535 1.1083002
#> [30,] -0.5989791  3.221706e-01 0.98523711 1.3060795
#> [31,] -0.7353867  1.301910e-01 1.08282286 1.6641028
#> [32,] -0.8606051 -5.265820e-02 0.70860378 1.2227227
#> [33,] -1.2552969 -5.013191e-01 0.56722050 0.8730172
#> [34,] -1.0955509 -2.438133e-01 0.87626236 1.2137720
#> [35,] -1.1780646 -6.001117e-01 0.26786681 0.6945999
#> [36,] -0.4035401  4.471785e-01 1.11289703 1.3855286
#> [37,] -0.5371455  1.904584e-01 0.99847959 1.2812803
#> [38,] -0.9167544 -6.232432e-02 0.66655278 0.9234413
#> [39,] -1.0596217 -1.480213e-01 0.49978356 0.7696024
#> [40,] -0.8677772 -3.069519e-01 0.38234927 0.8121058
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3902042 -0.1660597
#>  [2,] 0.3875414 -0.1614756
#>  [3,] 0.3635224 -0.1479128
#>  [4,] 0.3864509 -0.1515957
#>  [5,] 0.3763328 -0.1482039
#>  [6,] 0.3897354 -0.1647480
#>  [7,] 0.3734564 -0.1524528
#>  [8,] 0.3821627 -0.1496483
#>  [9,] 0.3895154 -0.1572628
#> [10,] 0.3967989 -0.1622677
#> [11,] 0.4015515 -0.1722534
#> [12,] 0.3459339 -0.1409649
#> [13,] 0.3826340 -0.1531681
#> [14,] 0.3912508 -0.1665278
#> [15,] 0.4027549 -0.1714747
#> [16,] 0.3971511 -0.1693882
#> [17,] 0.3697372 -0.1454770
#> [18,] 0.3693097 -0.1583588
#> [19,] 0.4003653 -0.1651853
#> [20,] 0.3596402 -0.1344475
#> [21,] 0.3727866 -0.1540604
#> [22,] 0.3857631 -0.1543354
#> [23,] 0.3838627 -0.1552592
#> [24,] 0.3847719 -0.1479758
#> [25,] 0.3797668 -0.1461925
#> [26,] 0.3821254 -0.1486453
#> [27,] 0.3712678 -0.1383540
#> [28,] 0.3878120 -0.1530055
#> [29,] 0.3903022 -0.1529155
#> [30,] 0.3944391 -0.1657612
#> [31,] 0.3964125 -0.1724544
#> [32,] 0.3833179 -0.1638378
#> [33,] 0.3732594 -0.1519268
#> [34,] 0.3812973 -0.1566313
#> [35,] 0.3650568 -0.1388711
#> [36,] 0.4024397 -0.1634605
#> [37,] 0.3955259 -0.1635193
#> [38,] 0.3817107 -0.1505922
#> [39,] 0.3678038 -0.1488354
#> [40,] 0.3633721 -0.1489241
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3644124 -0.3644124
#>  [2,]  0.3660925 -0.3660925
#>  [3,] -0.6251429  0.6251429
#>  [4,] -0.6313579  0.6313579
#>  [5,] -0.6282080  0.6282080
#>  [6,] -0.6351751  0.6351751
#>  [7,] -0.6285285  0.6285285
#>  [8,]  0.3700946 -0.3700946
#>  [9,]  0.3666122 -0.3666122
#> [10,]  0.3637635 -0.3637635
#> [11,] -0.6396407  0.6396407
#> [12,]  0.3806244 -0.3806244
#> [13,] -0.6308353  0.6308353
#> [14,] -0.6359384  0.6359384
#> [15,] -0.6397386  0.6397386
#> [16,]  0.3620357 -0.3620357
#> [17,] -0.6260280  0.6260280
#> [18,] -0.6289392  0.6289392
#> [19,]  0.3622641 -0.3622641
#> [20,] -0.6210689  0.6210689
#> [21,]  0.3712526 -0.3712526
#> [22,]  0.3681647 -0.3681647
#> [23,]  0.3683919 -0.3683919
#> [24,]  0.3698763 -0.3698763
#> [25,]  0.3714598 -0.3714598
#> [26,] -0.6296628  0.6296628
#> [27,]  0.3752822 -0.3752822
#> [28,]  0.3679974 -0.3679974
#> [29,] -0.6325606  0.6325606
#> [30,]  0.3635011 -0.3635011
#> [31,] -0.6385017  0.6385017
#> [32,]  0.3665246 -0.3665246
#> [33,]  0.3716403 -0.3716403
#> [34,] -0.6313304  0.6313304
#> [35,] -0.6233819  0.6233819
#> [36,] -0.6378166  0.6378166
#> [37,] -0.6362316  0.6362316
#> [38,] -0.6300201  0.6300201
#> [39,]  0.3736384 -0.3736384
#> [40,]  0.3746554 -0.3746554
#> 
#> $L
#> [1] 0.6901553 0.7506010 0.7461060 0.8058050 0.8102976 0.7293865
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5217391