Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13          V14
#>  [1,] -0.2242561 -0.21864219 -0.12493044 -0.10201189 -0.072627109 -0.173641967
#>  [2,] -0.3685877 -0.10609838 -0.14195010 -0.05731702  0.030433891  0.023622442
#>  [3,] -0.3768296 -0.13028679 -0.04042952 -0.18983714 -0.042485704 -0.147168286
#>  [4,] -0.4008061 -0.03490159 -0.08403715  0.02157603 -0.095202490 -0.052174185
#>  [5,] -0.1803339 -0.10857005 -0.02930542 -0.05849964 -0.127164120 -0.041882956
#>  [6,] -0.2082269 -0.09808111 -0.13063856 -0.03690011 -0.055348202 -0.156625825
#>  [7,] -0.3220535 -0.05450250 -0.04021634 -0.06841205 -0.017285737  0.074053204
#>  [8,] -0.3620971 -0.13640960 -0.04295473 -0.16070635  0.006111715  0.033670953
#>  [9,] -0.4008074 -0.26243116 -0.11170674 -0.12978032 -0.095709470 -0.163858292
#> [10,] -0.3451626 -0.06781899 -0.08615980 -0.04363310 -0.077799361 -0.005944309
#>               V15          V16         V17           V18         V19         V2
#>  [1,] -0.12402735 -0.119687696 -0.08077264 -0.0252595865 -0.04341657 -0.3086098
#>  [2,] -0.07830820  0.046496044 -0.06026173  0.0517661012  0.07734143 -0.2404212
#>  [3,] -0.06326004 -0.118267882 -0.05200145 -0.0251236651 -0.00847776 -0.3812652
#>  [4,] -0.06818831  0.086413278  0.06382198 -0.0219274598  0.12188064 -0.4107399
#>  [5,]  0.08843221  0.030845333  0.13475546  0.0572079557  0.05314076 -0.1934472
#>  [6,] -0.17501298 -0.200770654 -0.24658279 -0.1340966692 -0.19326281 -0.3110888
#>  [7,]  0.01631788  0.038193777  0.02982893  0.1097861918  0.15000197 -0.3513048
#>  [8,] -0.08905335  0.048383659  0.06202010 -0.0002557443 -0.01813636 -0.3189873
#>  [9,]  0.03191182  0.003976178 -0.07684275  0.0438313840  0.05166473 -0.2503863
#> [10,] -0.07463265  0.044568152  0.08372303 -0.0148318441 -0.06399445 -0.3261611
#>               V20         V21          V22         V23         V24         V25
#>  [1,]  0.04299043  0.04518652  0.020845114 -0.10888482 -0.08184669 -0.10682179
#>  [2,]  0.15062170  0.05735669  0.129610316  0.08633725  0.17388869  0.04187873
#>  [3,]  0.04148469  0.08966303  0.040528916  0.11391094  0.12817172  0.15494851
#>  [4,]  0.08391363  0.13865491  0.191667396  0.20550796  0.13805665  0.15986367
#>  [5,]  0.18372679  0.09726455  0.074571757  0.10017059  0.13638330  0.10856230
#>  [6,] -0.11068508 -0.12032256 -0.043523739 -0.03650664  0.07765462  0.15514451
#>  [7,]  0.16884212  0.19094296  0.098353850  0.20572111  0.12798671  0.20337477
#>  [8,]  0.07371839  0.17505487  0.124905997  0.03204677  0.17416801  0.05043641
#>  [9,]  0.12554709  0.07055873  0.167466516  0.10928828  0.17513105  0.16343016
#> [10,] -0.06038018  0.13183811  0.006314512 -0.05557335 -0.07090543 -0.03114827
#>               V26         V27         V28         V29         V3          V30
#>  [1,] 0.004761413 -0.14251250 -0.12415593 -0.11282453 -0.3149313 -0.093169669
#>  [2,] 0.066856687  0.03524147  0.19915640  0.05327762 -0.3566747  0.174976747
#>  [3,] 0.170073928  0.13123009  0.18801701  0.08937042 -0.2623892  0.003985044
#>  [4,] 0.138803378  0.18729872  0.09825937  0.18536600 -0.2836701  0.016213089
#>  [5,] 0.065341148  0.08552009  0.01581024  0.08536144 -0.2906949  0.052777397
#>  [6,] 0.122021509  0.03873907  0.08614689  0.11717261 -0.3515449 -0.084574748
#>  [7,] 0.106487582  0.22902191  0.16728734 -0.02401764 -0.2954622 -0.068725035
#>  [8,] 0.048580026  0.10595698 -0.01987320  0.02786542 -0.1944749  0.030760138
#>  [9,] 0.145290102  0.12561119  0.12465505 -0.05280268 -0.2628482  0.004090293
#> [10,] 0.098679344  0.09248116  0.05512456  0.07196740 -0.2924916  0.115449763
#>                V31           V32           V33          V34         V35
#>  [1,] -0.180055403 -0.1755592908 -0.0861667996 -0.006200932 -0.13060917
#>  [2,] -0.049654814 -0.0025042366 -0.0005403763 -0.089457491  0.05741469
#>  [3,] -0.135830628 -0.0442744652 -0.1158673380 -0.025213054  0.02196905
#>  [4,]  0.107526083  0.0902683902  0.0693918666  0.015327536 -0.09037875
#>  [5,] -0.050571976 -0.0804550576  0.0675698886  0.097622166  0.06850318
#>  [6,] -0.001798076  0.0002750213 -0.0925092259 -0.092357519 -0.20629058
#>  [7,] -0.049757268 -0.0222026819 -0.0228789805 -0.031982501 -0.07444395
#>  [8,]  0.012398220 -0.0773917749 -0.1184288596 -0.058723378 -0.17315273
#>  [9,]  0.037666170  0.0450867400 -0.0769520621 -0.138541452  0.01639102
#> [10,]  0.061406164 -0.0363537510 -0.0512686302 -0.027786767 -0.08005080
#>                V36          V37         V38          V39         V4
#>  [1,] -0.147225106 -0.050538953 -0.16148510 -0.127182634 -0.2883945
#>  [2,]  0.061865622  0.062742678  0.04907218 -0.006997717 -0.2468505
#>  [3,]  0.032275891 -0.110326163 -0.11450472 -0.024689164 -0.2587468
#>  [4,] -0.010153571 -0.009441764 -0.06354478  0.043586299 -0.2914071
#>  [5,] -0.001143187  0.063634115 -0.03850565 -0.094522002 -0.1870527
#>  [6,] -0.245018048 -0.177721409 -0.15403678 -0.017205258 -0.2026646
#>  [7,] -0.051197024 -0.098039332 -0.05226724 -0.043805048 -0.2604743
#>  [8,] -0.131954070 -0.183515081 -0.05854273 -0.105333537 -0.2075060
#>  [9,] -0.023742066 -0.007263178 -0.06194974 -0.130630753 -0.3013952
#> [10,] -0.043616052 -0.134133135 -0.02610092 -0.032008003 -0.2617875
#>                V40         V41          V42         V43         V44         V45
#>  [1,] -0.111792512 -0.10339885 -0.132672493 -0.17518323 -0.23246103 -0.14374877
#>  [2,] -0.033757665 -0.06201101 -0.061178053 -0.12305491 -0.13670487 -0.15740560
#>  [3,] -0.090526630 -0.08577268 -0.089443832 -0.20369904 -0.10290545 -0.18636310
#>  [4,]  0.046940245 -0.15361600  0.003410337 -0.09974030 -0.09191799 -0.10466877
#>  [5,] -0.053280678 -0.12351060 -0.086001013 -0.15684118 -0.12108435 -0.09182227
#>  [6,] -0.166397412 -0.11195180 -0.182386924 -0.12534410 -0.09921946 -0.16904885
#>  [7,] -0.065912176 -0.11519016 -0.051429446 -0.06520706 -0.13996519 -0.20453702
#>  [8,] -0.002687333 -0.04511084 -0.196792282  0.01447968 -0.20054480 -0.04018100
#>  [9,] -0.114289643 -0.10613138 -0.085822387 -0.06654429 -0.22671466 -0.16321334
#> [10,] -0.193919773 -0.06331176 -0.059457768 -0.11271389 -0.02747046 -0.08020896
#>               V46         V47        V48        V49         V5        V50
#>  [1,] -0.15715534 -0.16710374 -0.1273942 -0.2968678 -0.1424735 -0.2344055
#>  [2,] -0.13753768 -0.28015750 -0.2729332 -0.3890954 -0.3284897 -0.3578376
#>  [3,] -0.19997790 -0.22031672 -0.1761492 -0.2798424 -0.2052178 -0.2565189
#>  [4,] -0.11948808 -0.15651998 -0.3159741 -0.3302931 -0.2821592 -0.4085254
#>  [5,] -0.11909060 -0.16952781 -0.2089860 -0.1834193 -0.2604693 -0.3423981
#>  [6,] -0.07990943 -0.23614376 -0.2641609 -0.2356859 -0.1394999 -0.3710287
#>  [7,] -0.16343939 -0.13741845 -0.2091455 -0.2719310 -0.2993123 -0.2780826
#>  [8,] -0.22886303 -0.21901659 -0.2956534 -0.3133259 -0.3080588 -0.2308329
#>  [9,] -0.29045611 -0.25807997 -0.2225168 -0.2445363 -0.2088759 -0.3834241
#> [10,] -0.10875993 -0.09424134 -0.2884053 -0.3256675 -0.1595503 -0.3846388
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.1683688 -0.2972485 -0.3198915 -0.3305822 -0.2551383 -0.2592403
#>  [2,] -0.3727883 -0.4203142 -0.4118981 -0.4213038 -0.4167623 -0.3953001
#>  [3,] -0.3466736 -0.4132492 -0.4261100 -0.3609149 -0.3591549 -0.3443326
#>  [4,] -0.3903317 -0.3680025 -0.3564899 -0.4139586 -0.3224652 -0.4160248
#>  [5,] -0.2102792 -0.3221548 -0.2273766 -0.2287653 -0.2475304 -0.2766940
#>  [6,] -0.2747865 -0.2736094 -0.1938590 -0.3086803 -0.2694889 -0.2568823
#>  [7,] -0.3532074 -0.4846179 -0.4647872 -0.3610973 -0.3158301 -0.3570977
#>  [8,] -0.3608279 -0.2559019 -0.2943152 -0.3765713 -0.3205010 -0.4367844
#>  [9,] -0.3957358 -0.2952830 -0.2749857 -0.3449957 -0.3723545 -0.3435471
#> [10,] -0.3143151 -0.2229795 -0.2384359 -0.3451929 -0.3317686 -0.3031660
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.2373000 -0.2150051 -0.2658014 -0.1551549 -0.2298098 -0.15862011
#>  [2,] -0.3636221 -0.4693682 -0.4028975 -0.2074080 -0.4671524 -0.18311430
#>  [3,] -0.4057337 -0.2646343 -0.2840487 -0.2455528 -0.3773484 -0.28544716
#>  [4,] -0.4131998 -0.4085243 -0.3998038 -0.1770584 -0.4194201 -0.24853030
#>  [5,] -0.3404220 -0.2337323 -0.2115927 -0.2123589 -0.3295283 -0.14060973
#>  [6,] -0.2825873 -0.2527869 -0.2788872 -0.1767030 -0.3868090 -0.23916687
#>  [7,] -0.4674482 -0.3724857 -0.3842724 -0.2348431 -0.3603770 -0.18860002
#>  [8,] -0.2385995 -0.4132490 -0.2776414 -0.2739919 -0.2844956 -0.07765646
#>  [9,] -0.3272031 -0.3415807 -0.3443734 -0.1889144 -0.3069127 -0.17910608
#> [10,] -0.2597331 -0.3121774 -0.4142691 -0.2481406 -0.2701798 -0.14345585
#>               V8          V9
#>  [1,] -0.1202876 -0.06863517
#>  [2,] -0.2086608 -0.14490517
#>  [3,] -0.2575777 -0.14951865
#>  [4,] -0.2014390 -0.19519423
#>  [5,] -0.1544351 -0.07480603
#>  [6,] -0.1577078 -0.05542764
#>  [7,] -0.2860974 -0.13266251
#>  [8,] -0.1823279 -0.16270828
#>  [9,] -0.2059583 -0.08042622
#> [10,] -0.1117358 -0.20438448
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]       [,6]
#> [1,] 0.003360331 -0.08364876 -0.05887275 0.01980550 -0.12027654 0.04610495
#> [2,] 0.005641025 -0.02797625  0.07153560 0.09196655  0.03559402 0.04217503
#>            [,7]        [,8]         [,9]       [,10]
#> [1,] 0.06041708 0.010168491 -0.008450101 0.005869939
#> [2,] 0.04407780 0.005559667 -0.008575323 0.130238002
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 1.500306e-05 0.0001174049 8.711404e-05 1.029879e-04 0.0002192779
#>  [2,] 1.592202e-04 0.0012151214 1.331143e-03 1.458389e-03 0.0017173522
#>  [3,] 2.586306e-04 0.0019201666 2.281734e-03 2.398475e-03 0.0025013016
#>  [4,] 1.872349e-04 0.0013773768 1.561581e-03 1.675964e-03 0.0018591975
#>  [5,] 3.104660e-04 0.0022911570 2.766363e-03 2.866827e-03 0.0028945662
#>  [6,] 2.570542e-05 0.0001534070 7.847790e-05 8.717468e-05 0.0003201077
#>  [7,] 1.328056e-04 0.0009613903 9.833520e-04 1.157480e-03 0.0014329708
#>  [8,] 1.329094e-04 0.0009937386 1.025602e-03 1.144304e-03 0.0014652664
#>  [9,] 8.495052e-05 0.0006943949 6.073638e-04 7.460783e-04 0.0011418809
#> [10,] 2.627514e-04 0.0019412058 2.327978e-03 2.439465e-03 0.0024966606
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0003391671 0.0004474263 0.0006018724 0.0007363474 0.0008400398
#>  [2,] 0.0019291091 0.0022569607 0.0028650704 0.0034060004 0.0037043380
#>  [3,] 0.0025061240 0.0026078694 0.0031929666 0.0037329836 0.0039629193
#>  [4,] 0.0019684983 0.0021932745 0.0027283638 0.0031789659 0.0034195939
#>  [5,] 0.0027740087 0.0027575610 0.0032943287 0.0037712963 0.0039658347
#>  [6,] 0.0005760930 0.0008001248 0.0010865280 0.0013268158 0.0015055019
#>  [7,] 0.0016840168 0.0020703520 0.0026657484 0.0032078661 0.0035428185
#>  [8,] 0.0017955217 0.0022066390 0.0028544965 0.0034414601 0.0037945201
#>  [9,] 0.0015712231 0.0020943983 0.0027963549 0.0034280619 0.0038134375
#> [10,] 0.0024465401 0.0024844353 0.0030086645 0.0035122924 0.0037249636
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0007471473 2.651765e-05 0.0006938925 0.0007988976 0.0009742179
#>  [2,] 0.0036818676 2.622990e-04 0.0037063855 0.0042436800 0.0046062038
#>  [3,] 0.0044262935 3.956930e-04 0.0049968771 0.0057179845 0.0056733496
#>  [4,] 0.0035943056 2.967655e-04 0.0038063755 0.0043330981 0.0044444638
#>  [5,] 0.0047015317 4.607513e-04 0.0055652121 0.0063790158 0.0060882687
#>  [6,] 0.0012922946 4.312727e-05 0.0011958806 0.0013738500 0.0016993472
#>  [7,] 0.0033451178 2.147331e-04 0.0031591532 0.0035887828 0.0041358935
#>  [8,] 0.0036005920 2.110132e-04 0.0034930690 0.0039572344 0.0045203867
#>  [9,] 0.0033616196 1.631063e-04 0.0029722121 0.0033901086 0.0041564272
#> [10,] 0.0042587521 3.924918e-04 0.0049128987 0.0056360361 0.0055395478
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.001126593 0.001282969 0.001407975 0.001509380 0.001542416 0.001512529
#>  [2,] 0.004885825 0.005356081 0.005531098 0.005762379 0.005916160 0.005911619
#>  [3,] 0.005549160 0.005827668 0.005783971 0.006036220 0.006248993 0.006209897
#>  [4,] 0.004569823 0.004947763 0.005000086 0.005220702 0.005393236 0.005465267
#>  [5,] 0.005799478 0.006023286 0.005902643 0.006198748 0.006416275 0.006404239
#>  [6,] 0.001997447 0.002280557 0.002511095 0.002684266 0.002703513 0.002659812
#>  [7,] 0.004660180 0.005217379 0.005414120 0.005639602 0.005824132 0.005937039
#>  [8,] 0.005002279 0.005539643 0.005804968 0.006057838 0.006198668 0.006154645
#>  [9,] 0.004791145 0.005424306 0.005780975 0.006026637 0.006215762 0.006227431
#> [10,] 0.005366676 0.005603021 0.005540628 0.005789957 0.005986162 0.005918209
#>               V29           V3         V30         V31          V32         V33
#>  [1,] 0.001477992 3.209021e-05 0.001306988 0.001254673 0.0009935416 0.001039116
#>  [2,] 0.005625354 3.094215e-04 0.005105285 0.004677383 0.0038224119 0.003977938
#>  [3,] 0.005633859 4.646001e-04 0.005066236 0.004319106 0.0037182737 0.003766544
#>  [4,] 0.005123684 3.495265e-04 0.004657460 0.004134192 0.0034361222 0.003563086
#>  [5,] 0.005677013 5.427355e-04 0.005069109 0.004142703 0.0036629523 0.003631252
#>  [6,] 0.002633555 5.623770e-05 0.002375806 0.002278539 0.0018129762 0.001941005
#>  [7,] 0.005832838 2.530541e-04 0.005351310 0.005030822 0.0040646503 0.004276730
#>  [8,] 0.005930003 2.548135e-04 0.005392142 0.005057505 0.0041175675 0.004271429
#>  [9,] 0.006172154 1.958814e-04 0.005654984 0.005452034 0.0043049482 0.004565933
#> [10,] 0.005291261 4.607919e-04 0.004723461 0.003960450 0.0034592069 0.003459402
#>               V34         V35         V36         V37          V38          V39
#>  [1,] 0.001150682 0.001270481 0.001337759 0.001146861 0.0008064294 0.0005951076
#>  [2,] 0.004285579 0.004474338 0.004590073 0.004172052 0.0033258338 0.0026906401
#>  [3,] 0.003843026 0.003567699 0.003452073 0.003377835 0.0031563476 0.0027707637
#>  [4,] 0.003799672 0.003839993 0.003843638 0.003565292 0.0030508700 0.0025685989
#>  [5,] 0.003590690 0.003105780 0.002834661 0.002913892 0.0030647711 0.0028486401
#>  [6,] 0.002203080 0.002495312 0.002591329 0.002155962 0.0015042500 0.0010849045
#>  [7,] 0.004727560 0.005124178 0.005308724 0.004687195 0.0035962774 0.0028619372
#>  [8,] 0.004648640 0.005034946 0.005186836 0.004595758 0.0034784424 0.0027018059
#>  [9,] 0.005106431 0.005687652 0.006015062 0.005259560 0.0037566018 0.0027823549
#> [10,] 0.003469305 0.003141964 0.002970920 0.002964400 0.0029025899 0.0026066556
#>                 V4          V40          V41          V42          V43
#>  [1,] 4.261670e-05 0.0006461704 0.0005861194 0.0004578814 0.0003205901
#>  [2,] 3.530410e-04 0.0029134471 0.0027145905 0.0024108821 0.0018508356
#>  [3,] 5.141725e-04 0.0028570069 0.0027224360 0.0026878443 0.0022217511
#>  [4,] 3.945006e-04 0.0026982293 0.0025616433 0.0023796513 0.0018799137
#>  [5,] 6.033677e-04 0.0028428484 0.0027782475 0.0028948908 0.0024927518
#>  [6,] 7.484539e-05 0.0011369126 0.0010090495 0.0007608563 0.0005296723
#>  [7,] 2.952442e-04 0.0030629026 0.0027900611 0.0023768465 0.0018021699
#>  [8,] 3.047796e-04 0.0028994594 0.0026558143 0.0022465730 0.0017071160
#>  [9,] 2.350212e-04 0.0030986893 0.0027948870 0.0022197743 0.0015536278
#> [10,] 5.102197e-04 0.0026456971 0.0025266697 0.0025546936 0.0021673453
#>                V44          V45          V46          V47          V48
#>  [1,] 0.0002208657 0.0001104249 0.0001111801 0.0001181561 8.554933e-05
#>  [2,] 0.0014173143 0.0010835660 0.0009271284 0.0008328000 6.395137e-04
#>  [3,] 0.0018375367 0.0016759925 0.0013148485 0.0010843023 9.028299e-04
#>  [4,] 0.0014795619 0.0012459865 0.0010272051 0.0008734675 6.955472e-04
#>  [5,] 0.0021001863 0.0020350045 0.0015682719 0.0012410580 1.054358e-03
#>  [6,] 0.0003354030 0.0001134845 0.0001384927 0.0001605302 1.133561e-04
#>  [7,] 0.0013349474 0.0009067422 0.0008079408 0.0007339248 5.250451e-04
#>  [8,] 0.0012696113 0.0008669613 0.0007762952 0.0007096085 5.185484e-04
#>  [9,] 0.0010903771 0.0005850029 0.0005896343 0.0006295158 4.322314e-04
#> [10,] 0.0018142540 0.0016872262 0.0013131379 0.0010615378 8.894262e-04
#>                V49           V5          V50          V51          V52
#>  [1,] 3.256310e-05 7.100652e-05 2.748482e-05 1.055938e-05 1.172758e-05
#>  [2,] 3.425587e-04 5.229199e-04 1.496364e-04 9.706689e-05 8.171814e-05
#>  [3,] 5.236574e-04 7.085180e-04 1.805643e-04 1.448541e-04 1.167368e-04
#>  [4,] 3.878854e-04 5.549776e-04 1.503206e-04 1.113198e-04 9.142594e-05
#>  [5,] 6.232968e-04 8.110335e-04 1.974411e-04 1.745691e-04 1.386033e-04
#>  [6,] 3.407247e-05 1.166891e-04 4.627404e-05 1.402907e-05 1.891506e-05
#>  [7,] 2.638154e-04 4.621942e-04 1.384366e-04 8.251350e-05 7.201905e-05
#>  [8,] 2.556377e-04 4.625165e-04 1.374351e-04 8.100182e-05 7.282889e-05
#>  [9,] 1.881933e-04 3.792941e-04 1.297813e-04 5.865632e-05 5.553066e-05
#> [10,] 5.222046e-04 7.036767e-04 1.740388e-04 1.450942e-04 1.165888e-04
#>                V53          V54          V55          V56          V57
#>  [1,] 1.801186e-05 1.507983e-05 1.793380e-05 1.106972e-05 1.690521e-05
#>  [2,] 8.287893e-05 8.497373e-05 7.833781e-05 6.397337e-05 7.151233e-05
#>  [3,] 8.992198e-05 1.106199e-04 8.236343e-05 7.674708e-05 7.173410e-05
#>  [4,] 8.250424e-05 8.951455e-05 7.453333e-05 6.544984e-05 6.674642e-05
#>  [5,] 9.723559e-05 1.243160e-04 8.461612e-05 8.531482e-05 7.141234e-05
#>  [6,] 3.238588e-05 2.614475e-05 3.253820e-05 2.064951e-05 3.149619e-05
#>  [7,] 8.690261e-05 7.756179e-05 7.823142e-05 6.092076e-05 7.419910e-05
#>  [8,] 8.403685e-05 7.747341e-05 7.746165e-05 6.092732e-05 7.503408e-05
#>  [9,] 8.045672e-05 6.954873e-05 8.011314e-05 5.617896e-05 7.866906e-05
#> [10,] 8.665585e-05 1.077617e-04 7.707853e-05 7.457735e-05 6.642483e-05
#>                V58          V59           V6          V60           V7
#>  [1,] 1.025834e-05 1.084404e-05 0.0001422021 1.028548e-05 0.0001796677
#>  [2,] 5.876087e-05 6.357576e-05 0.0008156810 5.378563e-05 0.0009574480
#>  [3,] 7.052410e-05 7.580269e-05 0.0010018980 6.259341e-05 0.0011105017
#>  [4,] 6.058576e-05 6.604759e-05 0.0008195557 5.340299e-05 0.0009382754
#>  [5,] 7.775917e-05 8.410071e-05 0.0011018688 6.698579e-05 0.0011958117
#>  [6,] 1.942053e-05 2.059992e-05 0.0002323944 1.939507e-05 0.0003054330
#>  [7,] 5.757585e-05 6.252056e-05 0.0007664297 5.220824e-05 0.0009015917
#>  [8,] 5.595013e-05 5.954703e-05 0.0007716765 5.155391e-05 0.0009127603
#>  [9,] 5.136659e-05 5.603439e-05 0.0007023388 5.037812e-05 0.0008782130
#> [10,] 6.820566e-05 7.287090e-05 0.0009788218 5.978328e-05 0.0010682238
#>                 V8           V9
#>  [1,] 0.0001408191 0.0000958440
#>  [2,] 0.0009481729 0.0010392564
#>  [3,] 0.0012476247 0.0017317784
#>  [4,] 0.0009942937 0.0012206398
#>  [5,] 0.0014224780 0.0020998598
#>  [6,] 0.0002357972 0.0001224349
#>  [7,] 0.0008444399 0.0007800154
#>  [8,] 0.0008496765 0.0008299573
#>  [9,] 0.0007397325 0.0005337031
#> [10,] 0.0012267126 0.0017568654
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.01330767  0.03092817  0.03040599  0.02527034  0.02529057  0.01845349
#> [2,] -0.01330767 -0.03092817 -0.03040599 -0.02527034 -0.02529057 -0.01845349
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01944654  0.01897007  0.03130229  0.02580394
#> [2,] -0.01944654 -0.01897007 -0.03130229 -0.02580394
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.07840862 0.59185188 0.43562086 0.46929314 0.31695496 0.02893283
#>  [7] 0.51088188 0.38399635 0.28541702 0.39575392
#> 
#> $B[[2]]
#> [1] 0.09505750 0.03096559
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.002110260 0.008504872 0.008675455 0.007859801 0.008806955 0.003808516
#>  [7] 0.008627397 0.008800546 0.009007129 0.008176298
#> 
#> $vB[[2]]
#> [1]  0.02527955 -0.02527955
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#>  [2,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#>  [3,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#>  [4,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#>  [5,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#>  [6,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#>  [7,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#>  [8,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#>  [9,] 0.0093 0.0640 0.0888 0.1599 0.1541 0.2768 0.2176 0.2799 0.3491 0.2824
#> [10,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [11,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [12,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [13,] 0.0177 0.0613 0.1680 0.3476 0.4561 0.5188 0.6308 0.7201 0.5153 0.3818
#> [14,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [15,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [16,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [17,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [18,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [19,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [20,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [21,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [22,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [23,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [24,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [25,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [26,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [27,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [28,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [29,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [30,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [31,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [32,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [33,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [34,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [35,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [36,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [37,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [38,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [39,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [40,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#>  [2,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#>  [3,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#>  [4,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#>  [5,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#>  [6,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#>  [7,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#>  [8,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#>  [9,] 0.2479 0.0185 0.3005 0.4300 0.4684 0.4520 0.5026 0.6217 0.6571 0.6632
#> [10,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [11,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [12,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [13,] 0.2644 0.0300 0.3345 0.4865 0.6628 0.7389 0.9213 1.0000 0.7750 0.5593
#> [14,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [15,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [16,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [17,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [18,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [19,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [20,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [21,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [22,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [23,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [24,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [25,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [26,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [27,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [28,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [29,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [30,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [31,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [32,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [33,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [34,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [35,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [36,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [37,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [38,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [39,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [40,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#>  [2,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#>  [3,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#>  [4,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#>  [5,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#>  [6,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#>  [7,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#>  [8,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#>  [9,] 0.7321 0.8534 0.0056 1.0000 0.8448 0.6354 0.6308 0.6211 0.6976 0.5868
#> [10,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [11,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [12,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [13,] 0.6172 0.8635 0.0288 0.6592 0.4770 0.4983 0.3330 0.3076 0.2876 0.2226
#> [14,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [15,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [16,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [17,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [18,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [19,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [20,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [21,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [22,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [23,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [24,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [25,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [26,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [27,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [28,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [29,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [30,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [31,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [32,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [33,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [34,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [35,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [36,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [37,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [38,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [39,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [40,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#>  [2,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#>  [3,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#>  [4,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#>  [5,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#>  [6,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#>  [7,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#>  [8,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#>  [9,] 0.4889 0.3683 0.2043 0.0064 0.1469 0.2220 0.1449 0.1490 0.1211 0.1144
#> [10,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [11,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [12,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [13,] 0.0794 0.0603 0.1049 0.0394 0.0606 0.1530 0.0983 0.1643 0.1901 0.1107
#> [14,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [15,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [16,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [17,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [18,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [19,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [20,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [21,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [22,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [23,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [24,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [25,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [26,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [27,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [28,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [29,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [30,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [31,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [32,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [33,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [34,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [35,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [36,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [37,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [38,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [39,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [40,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#>  [2,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#>  [3,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#>  [4,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#>  [5,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#>  [6,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#>  [7,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#>  [8,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#>  [9,] 0.0791 0.0365 0.0152 0.0085 0.0260 0.0120 0.0022 0.0069 0.0064 0.0129
#> [10,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [11,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [12,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [13,] 0.1917 0.1467 0.0392 0.0356 0.0630 0.0270 0.0168 0.0102 0.0122 0.0044
#> [14,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [15,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [16,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [17,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [18,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [19,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [20,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [21,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [22,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [23,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [24,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [25,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [26,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [27,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [28,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [29,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [30,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [31,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [32,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [33,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [34,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [35,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [36,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [37,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [38,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [39,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [40,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#>  [2,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#>  [3,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#>  [4,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#>  [5,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#>  [6,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#>  [7,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#>  [8,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#>  [9,] 0.0114 0.0054 0.0089 0.0050 0.0058 0.0458 0.0025 0.0470 0.0057 0.0425
#> [10,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [11,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [12,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [13,] 0.0075 0.0124 0.0099 0.0057 0.0032 0.0526 0.0019 0.0688 0.0633 0.0624
#> [14,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [15,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [16,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [17,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [18,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [19,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [20,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [21,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [22,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [23,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [24,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [25,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [26,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [27,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [28,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [29,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [30,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [31,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [32,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [33,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [34,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [35,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [36,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [37,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [38,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [39,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [40,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2435605 0.7480170 0.5933243 0.7947344 0.7240204 0.2103857 0.7896627
#>  [2,] 0.1256517 0.6909328 0.4945404 0.7759951 0.6435915 0.1695931 0.6854236
#>  [3,] 0.1146210 0.6994865 0.5395191 0.7966732 0.6271234 0.2237399 0.7295805
#>  [4,] 0.1418066 0.6837700 0.5029579 0.7085641 0.6033188 0.2229797 0.5851179
#>  [5,] 0.2064595 0.7611196 0.6671956 0.8207632 0.6904204 0.3646712 0.7821668
#>  [6,] 0.1891100 0.7554883 0.6849484 0.8323871 0.6847870 0.4298657 0.7790277
#>  [7,] 0.1559927 0.6826433 0.4801768 0.7361620 0.6660877 0.1508344 0.7501131
#>  [8,] 0.2258370 0.7470462 0.6785674 0.7940222 0.6829342 0.4088060 0.7362918
#>  [9,] 0.1956985 0.7703928 0.6294076 0.8019076 0.7062764 0.3013240 0.7218054
#> [10,] 0.2200989 0.6454456 0.4579350 0.6692027 0.6512918 0.1529697 0.6911312
#> [11,] 0.1909415 0.7069382 0.6059420 0.7694845 0.6377230 0.3457052 0.7178932
#> [12,] 0.2142065 0.7615235 0.6900082 0.8323321 0.7077303 0.3811275 0.8081757
#> [13,] 0.1905870 0.7475534 0.6324817 0.8093066 0.7091918 0.3176085 0.7763545
#> [14,] 0.1638692 0.6881995 0.4553510 0.7262758 0.6526165 0.1600377 0.7149306
#> [15,] 0.2387456 0.7459446 0.6307566 0.7850098 0.7170707 0.2463530 0.7933598
#> [16,] 0.1888512 0.6711176 0.4762338 0.6552672 0.5817893 0.2000295 0.5349825
#> [17,] 0.1563070 0.7284178 0.5956602 0.8031240 0.6849504 0.2557032 0.7898227
#> [18,] 0.2009079 0.7626094 0.6518523 0.8055964 0.6792745 0.3796398 0.7052777
#> [19,] 0.2335346 0.8150097 0.7287283 0.8635254 0.7900263 0.3030329 0.8676211
#> [20,] 0.1644885 0.7132904 0.5216280 0.7730590 0.7050043 0.1624006 0.7883471
#> [21,] 0.2947167 0.7610984 0.6472201 0.8014076 0.7389574 0.2722699 0.8048144
#> [22,] 0.2092982 0.7702320 0.6952606 0.8369439 0.7186848 0.3697815 0.8164418
#> [23,] 0.2376478 0.7038263 0.6120499 0.7603783 0.6717880 0.2864510 0.7494419
#> [24,] 0.1548520 0.7505575 0.6104408 0.8144948 0.7349175 0.2013314 0.8116394
#> [25,] 0.2492597 0.7476850 0.6794063 0.7958114 0.6879892 0.3922935 0.7630158
#> [26,] 0.1373241 0.6178922 0.3870356 0.5689941 0.5727876 0.1078114 0.5196374
#> [27,] 0.1778009 0.7625853 0.6468158 0.8299044 0.7261309 0.2695509 0.8208086
#> [28,] 0.1218199 0.6889379 0.5311399 0.7815969 0.6205104 0.2233083 0.7177732
#> [29,] 0.1634362 0.7396581 0.6043783 0.8061302 0.6763920 0.2927925 0.7406865
#> [30,] 0.1732028 0.6830832 0.5596951 0.7428038 0.6182709 0.3049987 0.6804975
#> [31,] 0.2425347 0.7444655 0.5776344 0.7917409 0.7284198 0.2068179 0.7712593
#> [32,] 0.1892889 0.7831975 0.6591646 0.8201414 0.7194419 0.2906858 0.7886476
#> [33,] 0.1621301 0.7600814 0.6481153 0.8270923 0.7012438 0.3232281 0.7278192
#> [34,] 0.2102968 0.7981363 0.6879128 0.8662456 0.7905465 0.2539070 0.8656845
#> [35,] 0.1686372 0.7074974 0.5339264 0.7406860 0.6448711 0.2358694 0.7082965
#> [36,] 0.1797853 0.7165707 0.5864547 0.7702589 0.6532102 0.2932563 0.7105053
#> [37,] 0.1064318 0.6573207 0.4722860 0.7446440 0.6176472 0.1778788 0.6589393
#> [38,] 0.1622268 0.7879084 0.6647915 0.8155522 0.7144073 0.2465478 0.7668833
#> [39,] 0.2060301 0.7296478 0.6458011 0.7892393 0.6547433 0.3865061 0.7479475
#> [40,] 0.1603805 0.7293687 0.6021955 0.8117861 0.6802803 0.2772761 0.7833744
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6104555 0.6430315 0.5429246
#>  [2,] 0.4708086 0.4960936 0.4342457
#>  [3,] 0.5173009 0.5375353 0.4930768
#>  [4,] 0.3904100 0.4976395 0.4907165
#>  [5,] 0.6104551 0.6674402 0.5874683
#>  [6,] 0.6091730 0.6776214 0.5888436
#>  [7,] 0.5439245 0.5256451 0.4668369
#>  [8,] 0.5825165 0.6500354 0.6058283
#>  [9,] 0.5564052 0.6305120 0.5964941
#> [10,] 0.5111356 0.4925331 0.4749897
#> [11,] 0.5439391 0.5787715 0.5521385
#> [12,] 0.6391418 0.6800161 0.5834194
#> [13,] 0.6174889 0.6423213 0.5769169
#> [14,] 0.5167023 0.5103179 0.5137160
#> [15,] 0.6230976 0.6453528 0.5397260
#> [16,] 0.3841587 0.4479118 0.4763895
#> [17,] 0.5729712 0.6161808 0.5023907
#> [18,] 0.5509749 0.6221801 0.5879364
#> [19,] 0.6983255 0.7558363 0.6174704
#> [20,] 0.5895781 0.5779933 0.4965576
#> [21,] 0.6500150 0.6792078 0.5427506
#> [22,] 0.6464313 0.6876263 0.5846934
#> [23,] 0.5782222 0.6051726 0.5150763
#> [24,] 0.5978320 0.6385837 0.5179308
#> [25,] 0.6020159 0.6589043 0.5534742
#> [26,] 0.3363398 0.3633073 0.3848375
#> [27,] 0.6372340 0.6783674 0.5519810
#> [28,] 0.5059086 0.5177606 0.4967772
#> [29,] 0.5703790 0.6068038 0.5459567
#> [30,] 0.5127211 0.5320478 0.5403585
#> [31,] 0.6116138 0.6270438 0.5554055
#> [32,] 0.5965220 0.6724525 0.5704185
#> [33,] 0.5318967 0.6304782 0.5628953
#> [34,] 0.6928087 0.7456108 0.5916221
#> [35,] 0.5246586 0.5438733 0.5222090
#> [36,] 0.5648678 0.5979550 0.5447415
#> [37,] 0.4641895 0.5095693 0.4930397
#> [38,] 0.5424928 0.6350937 0.5390788
#> [39,] 0.5769193 0.6229527 0.5712914
#> [40,] 0.5760556 0.6087340 0.5202489
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5173707 0.4826293
#>  [2,] 0.5168342 0.4831658
#>  [3,] 0.5175554 0.4824446
#>  [4,] 0.5155027 0.4844973
#>  [5,] 0.5176675 0.4823325
#>  [6,] 0.5180726 0.4819274
#>  [7,] 0.5182274 0.4817726
#>  [8,] 0.5168214 0.4831786
#>  [9,] 0.5151627 0.4848373
#> [10,] 0.5178897 0.4821103
#> [11,] 0.5176828 0.4823172
#> [12,] 0.5179030 0.4820970
#> [13,] 0.5170346 0.4829654
#> [14,] 0.5173504 0.4826496
#> [15,] 0.5175695 0.4824305
#> [16,] 0.5155767 0.4844233
#> [17,] 0.5181334 0.4818666
#> [18,] 0.5160204 0.4839796
#> [19,] 0.5163194 0.4836806
#> [20,] 0.5177939 0.4822061
#> [21,] 0.5181665 0.4818335
#> [22,] 0.5176543 0.4823457
#> [23,] 0.5182912 0.4817088
#> [24,] 0.5166060 0.4833940
#> [25,] 0.5183050 0.4816950
#> [26,] 0.5161121 0.4838879
#> [27,] 0.5175250 0.4824750
#> [28,] 0.5173185 0.4826815
#> [29,] 0.5169860 0.4830140
#> [30,] 0.5172170 0.4827830
#> [31,] 0.5167771 0.4832229
#> [32,] 0.5165524 0.4834476
#> [33,] 0.5152698 0.4847302
#> [34,] 0.5165563 0.4834437
#> [35,] 0.5173071 0.4826929
#> [36,] 0.5175777 0.4824223
#> [37,] 0.5166954 0.4833046
#> [38,] 0.5151142 0.4848858
#> [39,] 0.5180877 0.4819123
#> [40,] 0.5178448 0.4821552
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]      [,2]        [,3]      [,4]      [,5]       [,6]
#>  [1,] -1.1332572 1.0880641  0.37772492 1.3537032 0.9644927 -1.3226023
#>  [2,] -1.9399646 0.8044838 -0.02183943 1.2424781 0.5909869 -1.5885135
#>  [3,] -2.0443850 0.8448539  0.15840689 1.3656303 0.5198959 -1.2440032
#>  [4,] -1.8003652 0.7711518  0.01183175 0.8884204 0.4193129 -1.2483855
#>  [5,] -1.3464003 1.1588276  0.69552817 1.5215271 0.8020855 -0.5551466
#>  [6,] -1.4558035 1.1281010  0.77660726 1.6026405 0.7758593 -0.2823990
#>  [7,] -1.6883518 0.7659462 -0.07933439 1.0261148 0.6905431 -1.7280719
#>  [8,] -1.2319687 1.0829202  0.74719593 1.3493431 0.7672890 -0.3689036
#>  [9,] -1.4133989 1.2105306  0.52967643 1.3982597 0.8773674 -0.8410010
#> [10,] -1.2650903 0.5990791 -0.16865885 0.7045812 0.6247224 -1.7114968
#> [11,] -1.4439041 0.8805596  0.43028625 1.2054028 0.5654951 -0.6379715
#> [12,] -1.2997534 1.1610504  0.80015763 1.6022462 0.8843860 -0.4847654
#> [13,] -1.4462005 1.0856058  0.54287799 1.4455106 0.8914621 -0.7647841
#> [14,] -1.6297167 0.7917153 -0.17907288 0.9758089 0.6305602 -1.6579479
#> [15,] -1.1595689 1.0770990  0.53546407 1.2951039 0.9299775 -1.1181585
#> [16,] -1.4574918 0.7132441 -0.09513651 0.6422736 0.3301232 -1.3861100
#> [17,] -1.6859664 0.9866102  0.38741466 1.4059347 0.7766164 -1.0684227
#> [18,] -1.3806294 1.1670393  0.62719118 1.4216462 0.7504396 -0.4910773
#> [19,] -1.1884595 1.4828967  0.98818007 1.8448850 1.3250842 -0.8328967
#> [20,] -1.6252037 0.9114190  0.08656587 1.2256651 0.8712432 -1.6404738
#> [21,] -0.8725849 1.1587109  0.60684188 1.3951153 1.0405565 -0.9831364
#> [22,] -1.3291610 1.2096217  0.82482977 1.6356630 0.9379470 -0.5331541
#> [23,] -1.1656188 0.8655853  0.45593705 1.1547548 0.7162832 -0.9126838
#> [24,] -1.6970421 1.1015876  0.44916566 1.4794849 1.0197174 -1.3779935
#> [25,] -1.1025647 1.0863032  0.75104474 1.3603181 0.7907353 -0.4376818
#> [26,] -1.8376956 0.4806112 -0.45978999 0.2777484 0.2932336 -2.1132937
#> [27,] -1.5313179 1.1669060  0.60507185 1.5849499 0.9750803 -0.9969026
#> [28,] -1.9753080 0.7951584  0.12472102 1.2749968 0.4917151 -1.2464903
#> [29,] -1.6328800 1.0441921  0.42374204 1.4250587 0.7372398 -0.8818601
#> [30,] -1.5630965 0.7679773  0.23992477 1.0605928 0.4822156 -0.8236062
#> [31,] -1.1388332 1.0693091  0.31307003 1.3354510 0.9866200 -1.3442141
#> [32,] -1.4546370 1.2843979  0.65957369 1.5173059 0.9416948 -0.8920557
#> [33,] -1.6424640 1.1531261  0.61076517 1.5651583 0.8532277 -0.7389762
#> [34,] -1.3231370 1.3746866  0.79037957 1.8681631 1.3282224 -1.0778824
#> [35,] -1.5953166 0.8832603  0.13591456 1.0495369 0.5965697 -1.1754603
#> [36,] -1.5178026 0.9275141  0.34932818 1.2097736 0.6331803 -0.8796209
#> [37,] -2.1277178 0.6513769 -0.11096959 1.0702477 0.4795736 -1.5307853
#> [38,] -1.6417517 1.3123634  0.68472060 1.4864990 0.9168866 -1.1171099
#> [39,] -1.3490234 0.9928365  0.60063299 1.3203461 0.6399549 -0.4620225
#> [40,] -1.6553996 0.9914218  0.41462131 1.4616580 0.7550603 -0.9580137
#>             [,7]        [,8]        [,9]        [,10]
#>  [1,] 1.32289328  0.44922712  0.58854610  0.172121988
#>  [2,] 0.77880988 -0.11689867 -0.01562590 -0.264549310
#>  [3,] 0.99249505  0.06923123  0.15042438 -0.027694745
#>  [4,] 0.34381901 -0.44558917 -0.00944211 -0.037138099
#>  [5,] 1.27833840  0.44922545  0.69662999  0.353509428
#>  [6,] 1.26000925  0.44383720  0.74286200  0.359186844
#>  [7,] 1.09921570  0.17615199  0.10267039 -0.132847565
#>  [8,] 1.02678322  0.33311233  0.61919483  0.429810163
#>  [9,] 0.95343493  0.22658542  0.53441388  0.390878107
#> [10,] 0.80541315  0.04454967 -0.02986971 -0.100124956
#> [11,] 0.93403498  0.17621081  0.31773249  0.209314723
#> [12,] 1.43819941  0.57164112  0.75384593  0.336826290
#> [13,] 1.24454704  0.47890327  0.58545374  0.310129445
#> [14,] 0.91945283  0.06683415  0.04127742  0.054877636
#> [15,] 1.34529801  0.50271700  0.59867350  0.159239745
#> [16,] 0.14015900 -0.47193341 -0.20911164 -0.094512470
#> [17,] 1.32385690  0.29398401  0.47336890  0.009562923
#> [18,] 0.87255815  0.20461061  0.49881177  0.355441248
#> [19,] 1.88008694  0.83933685  1.12998588  0.478825150
#> [20,] 1.31499075  0.36222173  0.31454090 -0.013769953
#> [21,] 1.41666061  0.61910498  0.75013343  0.171420756
#> [22,] 1.49242380  0.60338906  0.78904539  0.342070691
#> [23,] 1.09563820  0.31547940  0.42706513  0.060323457
#> [24,] 1.46069770  0.39643998  0.56922233  0.071754095
#> [25,] 1.16928529  0.41387176  0.65841543  0.214718153
#> [26,] 0.07858998 -0.67964833 -0.56103802 -0.469065349
#> [27,] 1.52183517  0.56337883  0.74627927  0.208678156
#> [28,] 0.93344250  0.02363544  0.07107231 -0.012891557
#> [29,] 1.04953980  0.28339750  0.43389662  0.184347054
#> [30,] 0.75605894  0.05089522  0.12836707  0.161785916
#> [31,] 1.21543543  0.45410086  0.51955550  0.222535899
#> [32,] 1.31679293  0.39099431  0.71929839  0.283558976
#> [33,] 0.98358598  0.12776011  0.53426900  0.252920774
#> [34,] 1.86332925  0.81328335  1.07533828  0.370675295
#> [35,] 0.88712493  0.09871462  0.17594565  0.088894487
#> [36,] 0.89783933  0.26094193  0.39695140  0.179445992
#> [37,] 0.65857079 -0.14348785  0.03828188 -0.027843149
#> [38,] 1.19079539  0.17038207  0.55413203  0.156634507
#> [39,] 1.08769551  0.31013927  0.50209980  0.287121844
#> [40,] 1.28544038  0.30660206  0.44199371  0.081039958
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.1614488 0.09193798
#>  [2,] 0.1502724 0.08291025
#>  [3,] 0.1620408 0.09179037
#>  [4,] 0.1463825 0.08435195
#>  [5,] 0.1737320 0.10303253
#>  [6,] 0.1784904 0.10616866
#>  [7,] 0.1554851 0.08254330
#>  [8,] 0.1725342 0.10522314
#>  [9,] 0.1607402 0.10007093
#> [10,] 0.1515084 0.07991917
#> [11,] 0.1693266 0.09856585
#> [12,] 0.1762023 0.10455952
#> [13,] 0.1686657 0.10050071
#> [14,] 0.1549898 0.08556015
#> [15,] 0.1635714 0.09326447
#> [16,] 0.1409791 0.07865219
#> [17,] 0.1649992 0.09243371
#> [18,] 0.1664458 0.10234247
#> [19,] 0.1704331 0.10513230
#> [20,] 0.1580657 0.08685986
#> [21,] 0.1661387 0.09344081
#> [22,] 0.1750336 0.10438696
#> [23,] 0.1655717 0.09237429
#> [24,] 0.1591142 0.09266564
#> [25,] 0.1726484 0.09939569
#> [26,] 0.1290262 0.06455534
#> [27,] 0.1676285 0.09749976
#> [28,] 0.1612665 0.09196496
#> [29,] 0.1647861 0.09681586
#> [30,] 0.1642851 0.09538989
#> [31,] 0.1598764 0.09274292
#> [32,] 0.1646355 0.09840190
#> [33,] 0.1619203 0.10082186
#> [34,] 0.1676311 0.10138170
#> [35,] 0.1588337 0.08957754
#> [36,] 0.1645893 0.09424940
#> [37,] 0.1538058 0.08699956
#> [38,] 0.1560389 0.09556364
#> [39,] 0.1741315 0.10174903
#> [40,] 0.1670281 0.09561840
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4826293 -0.4826293
#>  [2,]  0.4831658 -0.4831658
#>  [3,]  0.4824446 -0.4824446
#>  [4,] -0.5155027  0.5155027
#>  [5,] -0.5176675  0.5176675
#>  [6,]  0.4819274 -0.4819274
#>  [7,]  0.4817726 -0.4817726
#>  [8,] -0.5168214  0.5168214
#>  [9,] -0.5151627  0.5151627
#> [10,]  0.4821103 -0.4821103
#> [11,]  0.4823172 -0.4823172
#> [12,]  0.4820970 -0.4820970
#> [13,] -0.5170346  0.5170346
#> [14,] -0.5173504  0.5173504
#> [15,] -0.5175695  0.5175695
#> [16,] -0.5155767  0.5155767
#> [17,]  0.4818666 -0.4818666
#> [18,] -0.5160204  0.5160204
#> [19,] -0.5163194  0.5163194
#> [20,]  0.4822061 -0.4822061
#> [21,] -0.5181665  0.5181665
#> [22,]  0.4823457 -0.4823457
#> [23,]  0.4817088 -0.4817088
#> [24,]  0.4833940 -0.4833940
#> [25,] -0.5183050  0.5183050
#> [26,]  0.4838879 -0.4838879
#> [27,]  0.4824750 -0.4824750
#> [28,]  0.4826815 -0.4826815
#> [29,]  0.4830140 -0.4830140
#> [30,]  0.4827830 -0.4827830
#> [31,]  0.4832229 -0.4832229
#> [32,] -0.5165524  0.5165524
#> [33,] -0.5152698  0.5152698
#> [34,] -0.5165563  0.5165563
#> [35,] -0.5173071  0.5173071
#> [36,]  0.4824223 -0.4824223
#> [37,]  0.4833046 -0.4833046
#> [38,]  0.4848858 -0.4848858
#> [39,]  0.4819123 -0.4819123
#> [40,]  0.4821552 -0.4821552
#> 
#> $L
#> [1] 0.6920782 0.7107202 0.7030538 0.7640843 1.1093805 0.6862269
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174