Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12          V13
#>  [1,] -0.3757810 -0.03721528 -0.102942157 -0.084435238 -0.061560216
#>  [2,] -0.3421799 -0.10681870 -0.095825715  0.009444702 -0.004093584
#>  [3,] -0.2830605 -0.01809037 -0.023286220 -0.066142004  0.043225905
#>  [4,] -0.2804342 -0.15565328 -0.202956074 -0.039241154 -0.149756845
#>  [5,] -0.3873582 -0.06256617 -0.003428154 -0.103892843 -0.143409041
#>  [6,] -0.3536274 -0.17656631 -0.135915567 -0.021131833 -0.046918069
#>  [7,] -0.2906541 -0.21979661 -0.176012882 -0.123079360 -0.069826584
#>  [8,] -0.2136677 -0.12628858 -0.169514185 -0.214688069 -0.144494557
#>  [9,] -0.2474253 -0.18177751 -0.087204169 -0.023503840 -0.184009030
#> [10,] -0.3193712 -0.14319168  0.018164686  0.003421808  0.034932155
#>                V14          V15          V16          V17          V18
#>  [1,] -0.045067886 -0.124192815  0.034455105  0.024463542  0.141938835
#>  [2,] -0.009120275  0.032698748 -0.026514434 -0.029001813  0.087059163
#>  [3,]  0.033617890 -0.002067360 -0.085245331  0.052235578 -0.004354577
#>  [4,] -0.182024304 -0.143544118 -0.017083938 -0.069750173 -0.104685981
#>  [5,] -0.021274075 -0.038804187  0.037338066  0.007150592  0.020108170
#>  [6,] -0.097289694 -0.099680666  0.023081480 -0.012665388  0.017407647
#>  [7,] -0.058734908 -0.083752580 -0.088513761 -0.075123453 -0.144556348
#>  [8,] -0.061471733 -0.023268275 -0.002083468 -0.138145521  0.056439047
#>  [9,] -0.154818755  0.009425245 -0.108028731  0.059012817  0.075169709
#> [10,] -0.054082184 -0.047993419 -0.065113794 -0.026292736  0.011133075
#>               V19         V2          V20         V21         V22        V23
#>  [1,]  0.18807289 -0.3527436  0.095100892  0.14045889  0.12923201 0.12818550
#>  [2,]  0.02908386 -0.3406341  0.104197817  0.18192179  0.18206734 0.09227915
#>  [3,]  0.14699223 -0.2942532  0.134754416  0.06279649  0.16938582 0.11225730
#>  [4,]  0.04870172 -0.2085919 -0.066311654 -0.01363488  0.03008235 0.04535300
#>  [5,]  0.07344101 -0.3700699  0.028842165  0.14153878  0.25508045 0.06232188
#>  [6,] -0.02954242 -0.2595796 -0.009063234  0.06488750  0.05430127 0.21019430
#>  [7,] -0.22494414 -0.3522875  0.034335367 -0.02548135  0.03688819 0.10447651
#>  [8,] -0.03980236 -0.3776420  0.164069496  0.11821517  0.04394954 0.10856357
#>  [9,] -0.08977136 -0.2920234  0.030842347  0.10573743 -0.03722690 0.06857121
#> [10,]  0.16128757 -0.2832539  0.140819923  0.26587266  0.18664636 0.14202807
#>               V24        V25        V26        V27        V28         V29
#>  [1,]  0.10844855 0.17767716 0.18402134 0.09393653 0.20893971  0.08820261
#>  [2,]  0.10555931 0.02635386 0.03038625 0.08602579 0.04471634 -0.01328761
#>  [3,]  0.28171666 0.33854143 0.30054557 0.35673056 0.21271512  0.25939974
#>  [4,]  0.12635185 0.12111712 0.04504509 0.14226766 0.08522193  0.10733035
#>  [5,]  0.20785107 0.18657097 0.28332534 0.22216994 0.18672596  0.23432860
#>  [6,]  0.12112649 0.26382544 0.18452853 0.09988173 0.15821382  0.08097452
#>  [7,]  0.06815792 0.03265148 0.06023475 0.19583511 0.11436175  0.10584503
#>  [8,]  0.19522978 0.06456796 0.12643449 0.22067352 0.24684996  0.17836261
#>  [9,] -0.03870237 0.02137978 0.14511448 0.12095108 0.05938890  0.08686832
#> [10,]  0.22738280 0.16227479 0.18540179 0.22337440 0.15473888  0.21936509
#>               V3          V30          V31          V32         V33         V34
#>  [1,] -0.3831284 -0.001786463  0.107776436  0.046266061 -0.02823118  0.07839699
#>  [2,] -0.2225378  0.143582537  0.056276618 -0.080193918  0.04448790 -0.01328965
#>  [3,] -0.3279796  0.225427674  0.135509856  0.008713191  0.01816797  0.09904509
#>  [4,] -0.2383758  0.105314988 -0.103996173 -0.158899296 -0.13313789 -0.02184793
#>  [5,] -0.2179483  0.044819974  0.081040345  0.109396369  0.08442559  0.02806449
#>  [6,] -0.3787149 -0.006664622  0.045739277  0.038241480  0.04155687 -0.11396853
#>  [7,] -0.1844004  0.088101669 -0.005300513 -0.001401815 -0.05025599 -0.13518123
#>  [8,] -0.2937837  0.042265244 -0.102443467  0.018007067 -0.02264791 -0.03367841
#>  [9,] -0.3149400 -0.043037009 -0.063420848 -0.065876940 -0.01624068 -0.05197236
#> [10,] -0.3633060  0.150936819  0.016336805 -0.016608769  0.11073149  0.08674756
#>                V35          V36         V37          V38           V39
#>  [1,] -0.046546995  0.007585464 -0.04828466 -0.100600312  0.0353918482
#>  [2,]  0.092393848 -0.058182069 -0.05654013  0.086814347  0.0406804790
#>  [3,]  0.106010030  0.016670175  0.13517890  0.082522321  0.1002902736
#>  [4,] -0.113760086 -0.041247369 -0.08605021 -0.048082980 -0.1737630612
#>  [5,] -0.149245020 -0.022514653 -0.00833537 -0.003892287 -0.1382472174
#>  [6,]  0.002109388 -0.005080619 -0.06141957 -0.012616265 -0.1298008382
#>  [7,] -0.039458868 -0.153810259 -0.12358642 -0.255706516 -0.0756229305
#>  [8,]  0.040561137 -0.009877457  0.01427869 -0.032200174 -0.0717525546
#>  [9,] -0.241686063 -0.161446973 -0.12272409 -0.165535362 -0.0170777069
#> [10,]  0.061262090  0.017778806  0.07172912  0.121232324 -0.0001933105
#>               V4         V40         V41         V42           V43          V44
#>  [1,] -0.3326663 -0.09593782  0.03243945 -0.08591992 -0.1260075876 -0.103024965
#>  [2,] -0.2339633 -0.04240154 -0.07959611 -0.04023677 -0.0536181406  0.009157165
#>  [3,] -0.3061166  0.07048174 -0.11103981 -0.05876240 -0.0708243114 -0.099184300
#>  [4,] -0.2009848 -0.13033108 -0.17637730 -0.02314242 -0.0210669095 -0.167812723
#>  [5,] -0.2970954  0.01805394  0.02223822 -0.11221654 -0.0278531172 -0.093154838
#>  [6,] -0.3255050 -0.17677473 -0.17625117  0.01349116 -0.1501305905 -0.152699378
#>  [7,] -0.3202888 -0.01833370 -0.12788063 -0.10416940 -0.2145134597 -0.213233469
#>  [8,] -0.2407507 -0.13654738 -0.05194278 -0.08459340 -0.1333894610 -0.104481179
#>  [9,] -0.1745563 -0.02130937 -0.03288211 -0.03781106  0.0007415591 -0.062723769
#> [10,] -0.1808968  0.03478457 -0.04043392  0.02379193  0.0581534764 -0.069683706
#>               V45         V46        V47        V48        V49         V5
#>  [1,] -0.08354580 -0.16624542 -0.2429988 -0.2488730 -0.3444861 -0.1696252
#>  [2,] -0.11141911 -0.09496519 -0.1583389 -0.2633445 -0.3132989 -0.1822212
#>  [3,]  0.02840892 -0.13663625 -0.1046004 -0.1766386 -0.2606922 -0.2847203
#>  [4,] -0.22722493 -0.11439920 -0.1266880 -0.2322454 -0.1968646 -0.2387316
#>  [5,] -0.12440846 -0.19951046 -0.1643172 -0.2760485 -0.2535294 -0.1637972
#>  [6,] -0.06251029 -0.13110071 -0.2356758 -0.2202452 -0.2858995 -0.2566483
#>  [7,] -0.20752103 -0.22994056 -0.1367988 -0.2415994 -0.3392992 -0.2373135
#>  [8,] -0.14012310 -0.22561366 -0.1191133 -0.1359454 -0.3253623 -0.2466853
#>  [9,] -0.16801567 -0.12197639 -0.2588112 -0.1690591 -0.2761889 -0.1878471
#> [10,] -0.10349902 -0.13303820 -0.1840077 -0.1899502 -0.2594294 -0.2653069
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2712682 -0.3250796 -0.4134886 -0.4226086 -0.3104253 -0.4535754
#>  [2,] -0.2865155 -0.2953172 -0.2510373 -0.3300618 -0.2917422 -0.2712558
#>  [3,] -0.4174358 -0.4457849 -0.4215281 -0.3786392 -0.4529787 -0.4407642
#>  [4,] -0.3773505 -0.2958977 -0.3852707 -0.3296953 -0.2598423 -0.2936111
#>  [5,] -0.3367830 -0.2685994 -0.3062415 -0.3661779 -0.2427259 -0.3237274
#>  [6,] -0.4012194 -0.3605535 -0.3526538 -0.3339573 -0.3443888 -0.3387429
#>  [7,] -0.2716748 -0.2701635 -0.3496532 -0.3696723 -0.3819824 -0.3180569
#>  [8,] -0.2816990 -0.3760925 -0.3860916 -0.2643104 -0.2894864 -0.2728920
#>  [9,] -0.2555318 -0.2357596 -0.2783624 -0.2211703 -0.3881811 -0.2109192
#> [10,] -0.3347477 -0.3381166 -0.3396440 -0.3086659 -0.3330527 -0.2640382
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3563755 -0.2990204 -0.3039065 -0.3297774 -0.13552895 -0.3124747
#>  [2,] -0.3586771 -0.3692482 -0.3720698 -0.3596306 -0.08842562 -0.2743553
#>  [3,] -0.4612990 -0.2961344 -0.3585321 -0.3268160 -0.13760724 -0.4562544
#>  [4,] -0.3187302 -0.2149763 -0.2700718 -0.2715136 -0.26014416 -0.2899476
#>  [5,] -0.2437972 -0.4182430 -0.4029102 -0.3909348 -0.12030613 -0.4048537
#>  [6,] -0.3800261 -0.4259648 -0.4063541 -0.3650930 -0.32781266 -0.3611382
#>  [7,] -0.2390305 -0.3358913 -0.2537105 -0.3774914 -0.21140549 -0.2340049
#>  [8,] -0.3520830 -0.2788506 -0.3114971 -0.4102646 -0.17327599 -0.3277493
#>  [9,] -0.3158797 -0.2573375 -0.3993988 -0.3046355 -0.26797270 -0.3861326
#> [10,] -0.3378398 -0.4132263 -0.3742705 -0.3135984 -0.14906565 -0.3409978
#>                V7          V8          V9
#>  [1,] -0.09207177 -0.09204269 -0.17976281
#>  [2,] -0.08647858 -0.21689253 -0.08983514
#>  [3,] -0.18564021 -0.19249616 -0.05501517
#>  [4,] -0.25781107 -0.14941664 -0.13316320
#>  [5,] -0.25964065 -0.19380042 -0.10211256
#>  [6,] -0.17566641 -0.21575123 -0.16303881
#>  [7,] -0.24801486 -0.11190606 -0.23479510
#>  [8,] -0.23965098 -0.19455127 -0.25015797
#>  [9,] -0.17223914 -0.23869785 -0.07342738
#> [10,] -0.09415237 -0.15634688 -0.17148651
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]         [,4]       [,5]         [,6]
#> [1,] -0.04145297 0.04371013 -0.11527037 -0.017280494 0.03204722 -0.069843730
#> [2,]  0.06647513 0.09082170  0.02698534 -0.005246821 0.06246316 -0.008902766
#>             [,7]        [,8]         [,9]      [,10]
#> [1,] -0.19174644 -0.02912935 -0.052687370 0.04952046
#> [2,]  0.07392594  0.08724159 -0.007180999 0.05606304
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10           V11           V12          V13
#>  [1,] 5.122089e-05 0.0002888282  2.171374e-04  2.116589e-04 0.0003999627
#>  [2,] 5.656232e-05 0.0003316110  1.249430e-04  9.004080e-05 0.0004803777
#>  [3,] 3.436191e-05 0.0001751501  1.746162e-04  1.741015e-04 0.0002242706
#>  [4,] 2.889019e-05 0.0001585610 -1.513212e-04 -2.317863e-04 0.0002657292
#>  [5,] 3.065207e-05 0.0001824216 -3.232452e-05 -4.328177e-05 0.0003067343
#>  [6,] 3.866492e-05 0.0002449018  4.584123e-06 -2.635650e-05 0.0003695054
#>  [7,] 1.406649e-05 0.0001813045  6.116754e-04  7.508214e-04 0.0001263499
#>  [8,] 5.136352e-05 0.0003221405  2.415529e-04  2.441545e-04 0.0004353165
#>  [9,] 3.123153e-05 0.0002047501  2.951159e-05 -8.566100e-06 0.0002879858
#> [10,] 2.865948e-05 0.0001965422  2.509465e-05 -1.758555e-05 0.0002666303
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0007784747  0.0009067192  0.0009891321  0.0010619650  0.0012703017
#>  [2,]  0.0012009266  0.0014645590  0.0016420690  0.0017789289  0.0021098648
#>  [3,]  0.0003438706  0.0003796862  0.0004078402  0.0004202584  0.0005201801
#>  [4,]  0.0010869259  0.0014003647  0.0016128008  0.0017137922  0.0020091949
#>  [5,]  0.0010035503  0.0012771824  0.0014101805  0.0015045019  0.0018303866
#>  [6,]  0.0011086560  0.0014132317  0.0015922054  0.0016883061  0.0020074522
#>  [7,] -0.0008548769 -0.0012428166 -0.0014168767 -0.0014677546 -0.0017075981
#>  [8,]  0.0007995793  0.0009324801  0.0010579398  0.0011420669  0.0013516802
#>  [9,]  0.0007890668  0.0009766329  0.0011228953  0.0011967439  0.0014140030
#> [10,]  0.0008434521  0.0010927124  0.0011804883  0.0012253820  0.0014405828
#>                 V19           V2           V20           V21           V22
#>  [1,]  0.0011131734 5.580398e-05  0.0012155421  0.0015013828  0.0020095344
#>  [2,]  0.0017135522 5.800052e-05  0.0017971219  0.0022541147  0.0030858068
#>  [3,]  0.0005496998 3.840248e-05  0.0006544388  0.0007737414  0.0009528255
#>  [4,]  0.0015045872 1.890759e-05  0.0015341243  0.0019369281  0.0026953585
#>  [5,]  0.0013527803 2.881641e-05  0.0013317220  0.0017299811  0.0025610772
#>  [6,]  0.0015405453 4.146380e-05  0.0015683681  0.0019843243  0.0027835129
#>  [7,] -0.0009515493 3.892078e-05 -0.0008677876 -0.0012112780 -0.0019800881
#>  [8,]  0.0012141718 5.893351e-05  0.0013228845  0.0016040492  0.0020647621
#>  [9,]  0.0011559479 3.173197e-05  0.0012293857  0.0015153708  0.0020150626
#> [10,]  0.0010684469 3.300339e-05  0.0010732743  0.0014050554  0.0020469887
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0019850048  0.0020346373  0.0022092384  0.0022119146  0.0022036442
#>  [2,]  0.0030765199  0.0031616814  0.0035217905  0.0035813638  0.0034917076
#>  [3,]  0.0009002023  0.0008790297  0.0009082194  0.0008915672  0.0009279656
#>  [4,]  0.0027315864  0.0028423798  0.0033050505  0.0034579739  0.0033012283
#>  [5,]  0.0025298519  0.0026185100  0.0029942289  0.0030039475  0.0029013918
#>  [6,]  0.0027362920  0.0028113802  0.0032010607  0.0032636705  0.0031636317
#>  [7,] -0.0020356174 -0.0021195271 -0.0026589340 -0.0028595087 -0.0025593765
#>  [8,]  0.0020341781  0.0020797812  0.0022598213  0.0022804007  0.0022786684
#>  [9,]  0.0020138603  0.0020690337  0.0023478241  0.0024467682  0.0023838807
#> [10,]  0.0020147236  0.0020723749  0.0023348511  0.0023575280  0.0022602877
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0021256976  0.0019465500  6.695422e-05  0.0018638703  0.0019507730
#>  [2,]  0.0032874794  0.0030198278  9.163305e-05  0.0029358969  0.0031730136
#>  [3,]  0.0009260664  0.0008229666  3.441208e-05  0.0007808707  0.0007941563
#>  [4,]  0.0030148444  0.0027692416  6.741939e-05  0.0027429779  0.0030847618
#>  [5,]  0.0027144856  0.0024844844  5.359564e-05  0.0024039261  0.0027500113
#>  [6,]  0.0029655261  0.0027148780  7.128680e-05  0.0026671067  0.0029810818
#>  [7,] -0.0021770728 -0.0020065521 -2.705629e-05 -0.0020783909 -0.0026487572
#>  [8,]  0.0022078989  0.0020235605  6.985794e-05  0.0019591855  0.0020236636
#>  [9,]  0.0022214977  0.0020297405  5.827987e-05  0.0019883016  0.0021573813
#> [10,]  0.0020971016  0.0018880211  5.760849e-05  0.0018211607  0.0020865182
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0015614697  0.0014409112  0.0013699156  0.0015739411  0.0016865300
#>  [2,]  0.0025463727  0.0023928178  0.0022939514  0.0026714476  0.0029267092
#>  [3,]  0.0006498862  0.0005728447  0.0005373611  0.0005831146  0.0005703588
#>  [4,]  0.0024828475  0.0023998869  0.0023564063  0.0027619617  0.0030505457
#>  [5,]  0.0021329217  0.0019472687  0.0018282475  0.0022456097  0.0025684926
#>  [6,]  0.0023453879  0.0021889837  0.0021066030  0.0025197715  0.0028178983
#>  [7,] -0.0021348396 -0.0020913737 -0.0020884919 -0.0026357404 -0.0030545734
#>  [8,]  0.0016298208  0.0015188462  0.0014614447  0.0016363017  0.0017314901
#>  [9,]  0.0017444006  0.0016651732  0.0016254194  0.0018542377  0.0020126841
#> [10,]  0.0016135777  0.0014455169  0.0013422624  0.0016771410  0.0019512890
#>                V37           V38           V39           V4           V40
#>  [1,]  0.001367209  0.0009678610  0.0008395308 5.640656e-05  0.0010084180
#>  [2,]  0.002355050  0.0015797321  0.0013072208 5.778036e-05  0.0016448293
#>  [3,]  0.000457712  0.0003752491  0.0003438810 4.031237e-05  0.0003875927
#>  [4,]  0.002403852  0.0015276443  0.0011885477 1.489423e-05  0.0015773710
#>  [5,]  0.002010432  0.0012358460  0.0010510963 1.024688e-05  0.0014825175
#>  [6,]  0.002235339  0.0014314051  0.0011882059 2.877207e-05  0.0015879679
#>  [7,] -0.002305271 -0.0012696088 -0.0008740238 5.343864e-05 -0.0013816443
#>  [8,]  0.001414021  0.0010306741  0.0008939792 6.152262e-05  0.0010486591
#>  [9,]  0.001615873  0.0010878154  0.0008831513 3.310176e-05  0.0011035666
#> [10,]  0.001537125  0.0009297761  0.0008061345 2.542501e-05  0.0011452459
#>                 V41          V42          V43           V44           V45
#>  [1,]  0.0006911748 0.0003768332 3.062095e-04  1.540674e-04  1.423277e-04
#>  [2,]  0.0010437924 0.0003888711 2.639282e-04  4.047078e-05 -1.872747e-05
#>  [3,]  0.0003042277 0.0002375608 2.182192e-04  1.526912e-04  1.697748e-04
#>  [4,]  0.0009421899 0.0001504207 1.640116e-05 -1.758650e-04 -2.768830e-04
#>  [5,]  0.0008518495 0.0001602294 5.679172e-05 -1.338725e-04 -2.205902e-04
#>  [6,]  0.0009501688 0.0002197281 9.659231e-05 -1.036427e-04 -1.712907e-04
#>  [7,] -0.0006574597 0.0003219297 4.529226e-04  6.004188e-04  7.638952e-04
#>  [8,]  0.0007312042 0.0004078240 3.349779e-04  1.894230e-04  1.846376e-04
#>  [9,]  0.0007021282 0.0002228310 1.339170e-04 -9.709195e-07 -4.812495e-05
#> [10,]  0.0006659613 0.0001539708 6.176000e-05 -7.495964e-05 -1.523692e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  1.751744e-04  1.626603e-04  1.007130e-04  3.472678e-05  6.337251e-05
#>  [2,]  5.971473e-05  1.226720e-04  8.459928e-05  5.461352e-06  5.208141e-05
#>  [3,]  1.674797e-04  1.286849e-04  7.787747e-05  3.841252e-05  4.741702e-05
#>  [4,] -1.788185e-04 -3.844560e-05 -9.871385e-07 -5.270524e-05 -1.068110e-05
#>  [5,] -1.066280e-04  1.268201e-05 -4.023298e-06 -3.961763e-05  8.619913e-06
#>  [6,] -5.916444e-05  4.879145e-05  2.955761e-05 -2.641943e-05  1.946346e-05
#>  [7,]  6.192906e-04  3.380427e-04  1.896722e-04  1.683003e-04  9.650414e-05
#>  [8,]  2.093164e-04  1.886485e-04  1.233789e-04  4.792579e-05  6.395854e-05
#>  [9,]  1.040786e-06  5.729607e-05  5.094214e-05 -3.295949e-06  2.195101e-05
#> [10,] -8.585963e-05  1.453150e-05  6.347123e-06 -2.456613e-05  3.659927e-05
#>                 V50           V51          V52           V53           V54
#>  [1,]  4.959676e-05  1.990472e-05 3.125515e-05  2.760411e-05  1.657881e-05
#>  [2,]  6.747083e-05  1.525530e-05 3.445906e-05  4.178820e-05  2.313464e-05
#>  [3,]  2.738913e-05  1.651755e-05 2.175540e-05  1.399701e-05  9.885905e-06
#>  [4,]  5.388439e-05 -3.053006e-06 1.561845e-05  3.913562e-05  2.095802e-05
#>  [5,]  5.237508e-05 -2.577540e-06 2.050821e-05  3.110376e-05  1.086154e-05
#>  [6,]  5.961630e-05  4.729736e-06 2.545873e-05  3.574870e-05  1.726186e-05
#>  [7,] -2.073124e-05  3.873604e-05 1.343559e-05 -2.972565e-05 -1.085357e-05
#>  [8,]  5.421688e-05  2.378730e-05 3.257542e-05  2.841744e-05  1.875236e-05
#>  [9,]  4.458821e-05  8.603220e-06 1.938397e-05  2.885104e-05  1.730117e-05
#> [10,]  3.611599e-05 -1.595405e-06 1.696293e-05  2.486873e-05  8.500779e-06
#>                 V55           V56           V57           V58           V59
#>  [1,]  1.494920e-05  1.255701e-05  2.179910e-05  6.330654e-06  1.095968e-05
#>  [2,]  1.865402e-05  1.974287e-05  3.544326e-05  5.376323e-06  1.516995e-05
#>  [3,]  9.261725e-06  6.186120e-06  8.765404e-06  4.972726e-06  5.962178e-06
#>  [4,]  1.454434e-05  1.922731e-05  3.561673e-05  5.394779e-07  1.242441e-05
#>  [5,]  1.000909e-05  1.071289e-05  2.933560e-05 -1.802753e-06  5.146569e-06
#>  [6,]  1.356984e-05  1.552153e-05  3.333004e-05  8.784112e-07  9.713909e-06
#>  [7,] -3.932345e-06 -1.341241e-05 -3.136673e-05  9.154460e-06 -4.525050e-06
#>  [8,]  1.589600e-05  1.437874e-05  2.277699e-05  7.174464e-06  1.220908e-05
#>  [9,]  1.256375e-05  1.483773e-05  2.439667e-05  3.190492e-06  1.047028e-05
#> [10,]  7.555457e-06  9.593748e-06  2.294100e-05 -8.136418e-07  5.825100e-06
#>                  V6           V60            V7            V8           V9
#>  [1,]  0.0002172726  1.983104e-05  0.0002707709  0.0002579205 0.0002201547
#>  [2,]  0.0003140956  3.195926e-05  0.0004348137  0.0003646685 0.0002400251
#>  [3,]  0.0001044079  8.507192e-06  0.0001120755  0.0001235201 0.0001439557
#>  [4,]  0.0002621715  3.099672e-05  0.0004210769  0.0002933581 0.0001059307
#>  [5,]  0.0002575744  2.637805e-05  0.0003645680  0.0002680157 0.0000815418
#>  [6,]  0.0002880300  2.998245e-05  0.0004125072  0.0003199044 0.0001538418
#>  [7,] -0.0001570079 -2.615470e-05 -0.0003324512 -0.0001406791 0.0001842175
#>  [8,]  0.0002281784  2.108133e-05  0.0002957913  0.0002800860 0.0002560331
#>  [9,]  0.0002058755  2.203001e-05  0.0003018159  0.0002398858 0.0001569312
#> [10,]  0.0002236418  2.001951e-05  0.0002900738  0.0002391399 0.0001172451
#> 
#> $vW[[2]]
#>            [,1]        [,2]       [,3]        [,4]       [,5]        [,6]
#> [1,]  0.1047313  0.08915342  0.1144208  0.06408836  0.1061399  0.09718731
#> [2,] -0.1047313 -0.08915342 -0.1144208 -0.06408836 -0.1061399 -0.09718731
#>             [,7]        [,8]        [,9]      [,10]
#> [1,]  0.06574039  0.09286845  0.06029333  0.1098783
#> [2,] -0.06574039 -0.09286845 -0.06029333 -0.1098783
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6558304 0.3109086 0.6965411 0.1628167 0.4619709 0.5679036 0.2372755
#>  [8] 0.2779830 0.1556417 0.7190415
#> 
#> $B[[2]]
#> [1] -0.1408864  0.1261291
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.002852538  0.004592247  0.001194728  0.004434712  0.003830550
#>  [6]  0.004226017 -0.003629121  0.002980459  0.003154053  0.003005847
#> 
#> $vB[[2]]
#> [1]  0.1214571 -0.1214571
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#>  [2,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#>  [3,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#>  [4,] 0.0223 0.1487 0.1156 0.1654 0.3833 0.3598 0.1713 0.1136 0.0349 0.3796
#>  [5,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#>  [6,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#>  [7,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#>  [8,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#>  [9,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#> [10,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#> [11,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [12,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [13,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#> [14,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [15,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [16,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [17,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [18,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [19,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [20,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [21,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [22,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [23,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [24,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [25,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [26,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [27,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [28,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [29,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [30,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [31,] 0.0454 0.1675 0.2799 0.3323 0.4012 0.4296 0.5350 0.5411 0.6870 0.8045
#> [32,] 0.0195 0.0242 0.0728 0.0639 0.3002 0.3854 0.4767 0.4602 0.3175 0.4160
#> [33,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [34,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [35,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [36,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [37,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [38,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [39,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [40,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#>  [2,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#>  [3,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#>  [4,] 0.7401 0.0375 0.9925 0.9802 0.8890 0.6712 0.4286 0.3374 0.7366 0.9611
#>  [5,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#>  [6,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#>  [7,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#>  [8,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#>  [9,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#> [10,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#> [11,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [12,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [13,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#> [14,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [15,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [16,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [17,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [18,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [19,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [20,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [21,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [22,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [23,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [24,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [25,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [26,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [27,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [28,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [29,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [30,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [31,] 0.9194 0.0472 0.9169 1.0000 0.9972 0.9093 0.7918 0.6705 0.5324 0.3572
#> [32,] 0.6428 0.0213 1.0000 0.8631 0.5212 0.3156 0.5952 0.7732 0.6042 0.4375
#> [33,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [34,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [35,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [36,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [37,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [38,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [39,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [40,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#>  [2,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#>  [3,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#>  [4,] 0.7353 0.4856 0.0484 0.1594 0.3007 0.4096 0.3170 0.3305 0.3408 0.2186
#>  [5,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#>  [6,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#>  [7,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#>  [8,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#>  [9,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#> [10,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#> [11,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [12,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [13,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#> [14,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [15,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [16,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [17,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [18,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [19,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [20,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [21,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [22,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [23,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [24,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [25,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [26,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [27,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [28,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [29,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [30,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [31,] 0.2484 0.3161 0.0697 0.3775 0.3138 0.1713 0.2937 0.5234 0.5926 0.5437
#> [32,] 0.5487 0.4720 0.0058 0.6235 0.3851 0.1590 0.3891 0.5294 0.3504 0.4480
#> [33,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [34,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [35,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [36,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [37,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [38,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [39,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [40,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#>  [2,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#>  [3,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#>  [4,] 0.2463 0.2726 0.1680 0.0475 0.2792 0.2558 0.1740 0.2121 0.1099 0.0985
#>  [5,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#>  [6,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#>  [7,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#>  [8,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#>  [9,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#> [10,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#> [11,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [12,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [13,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#> [14,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [15,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [16,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [17,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [18,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [19,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [20,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [21,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [22,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [23,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [24,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [25,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [26,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [27,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [28,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [29,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [30,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [31,] 0.4516 0.3379 0.3215 0.1021 0.2178 0.1674 0.2634 0.2980 0.2037 0.1155
#> [32,] 0.4041 0.5031 0.6475 0.0190 0.5493 0.3548 0.2028 0.1882 0.0845 0.1315
#> [33,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [34,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [35,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [36,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [37,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [38,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [39,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [40,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#>  [2,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#>  [3,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#>  [4,] 0.1271 0.1459 0.1164 0.0777 0.0647 0.0439 0.0061 0.0145 0.0128 0.0145
#>  [5,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#>  [6,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#>  [7,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#>  [8,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#>  [9,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#> [10,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#> [11,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [12,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [13,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#> [14,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [15,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [16,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [17,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [18,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [19,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [20,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [21,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [22,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [23,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [24,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [25,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [26,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [27,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [28,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [29,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [30,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [31,] 0.0919 0.0882 0.0228 0.0380 0.1397 0.0142 0.0137 0.0120 0.0042 0.0238
#> [32,] 0.1590 0.0562 0.0617 0.0343 0.0319 0.0370 0.0261 0.0157 0.0074 0.0271
#> [33,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [34,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [35,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [36,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [37,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [38,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [39,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [40,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#>  [2,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#>  [3,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#>  [4,] 0.0058 0.0049 0.0065 0.0093 0.0059 0.0591 0.0022 0.0753 0.0098 0.0684
#>  [5,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#>  [6,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#>  [7,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#>  [8,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#>  [9,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> [10,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#> [11,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [12,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [13,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#> [14,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [15,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [16,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [17,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [18,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [19,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [20,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [21,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [22,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [23,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [24,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [25,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [26,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [27,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [28,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [29,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [30,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [31,] 0.0129 0.0084 0.0218 0.0321 0.0154 0.1493 0.0053 0.1487 0.0771 0.1171
#> [32,] 0.0203 0.0089 0.0095 0.0095 0.0021 0.0571 0.0053 0.1004 0.0668 0.0691
#> [33,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [34,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [35,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [36,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [37,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [38,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [39,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [40,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7943781 0.6952543 0.9072116 0.3545060 0.7835471 0.6635803 0.3099934
#>  [2,] 0.8149071 0.7050156 0.9340243 0.4509284 0.8366098 0.7223938 0.4409377
#>  [3,] 0.8260987 0.7176942 0.9306408 0.5008591 0.8402549 0.7351587 0.4923932
#>  [4,] 0.8081726 0.7013243 0.9081360 0.4407951 0.8146416 0.6901874 0.4397673
#>  [5,] 0.8107404 0.7033427 0.9053702 0.3789057 0.7918354 0.6766587 0.3499365
#>  [6,] 0.8291563 0.7013202 0.9366073 0.4050725 0.8334294 0.7052077 0.3559559
#>  [7,] 0.8364230 0.7198068 0.9330344 0.4535675 0.8374097 0.7216883 0.3990113
#>  [8,] 0.7656180 0.6402856 0.9153083 0.3830946 0.7910515 0.6568400 0.3423231
#>  [9,] 0.8549252 0.7321586 0.9570419 0.5259037 0.8820254 0.7935084 0.5219491
#> [10,] 0.7749921 0.6646410 0.9126474 0.4397301 0.8095314 0.6784985 0.4228037
#> [11,] 0.8366413 0.7142606 0.9432162 0.5296799 0.8642703 0.7698880 0.5042927
#> [12,] 0.7353304 0.6249426 0.8958327 0.3356919 0.7546467 0.5958276 0.3625655
#> [13,] 0.6097983 0.5706435 0.8304416 0.2239451 0.5912288 0.4179327 0.1900803
#> [14,] 0.7507368 0.6551586 0.8716484 0.3122712 0.7323450 0.5880582 0.2713430
#> [15,] 0.8223691 0.6989964 0.9572317 0.3677832 0.8306107 0.7021394 0.3163689
#> [16,] 0.8475578 0.7226619 0.9399014 0.4704156 0.8629480 0.7509986 0.4340869
#> [17,] 0.7672501 0.6479214 0.9196577 0.3473531 0.7882640 0.6151761 0.3048912
#> [18,] 0.8036513 0.6885076 0.9242674 0.4416223 0.8192742 0.7240646 0.4092806
#> [19,] 0.8594761 0.7405628 0.9331945 0.4850359 0.8631878 0.7654021 0.4615958
#> [20,] 0.7464643 0.6371762 0.8767517 0.3177956 0.7705543 0.6109935 0.3338004
#> [21,] 0.7880043 0.6733890 0.9056681 0.3297343 0.7862197 0.6197920 0.3014522
#> [22,] 0.8338619 0.7024894 0.9441710 0.5255476 0.8618123 0.7688478 0.5151053
#> [23,] 0.8362752 0.7024486 0.9414077 0.4978757 0.8533411 0.7496623 0.4631057
#> [24,] 0.8707614 0.7477465 0.9444402 0.4796310 0.8794097 0.7781269 0.4474064
#> [25,] 0.7993560 0.6793391 0.9340728 0.4861563 0.8339820 0.7374734 0.5129186
#> [26,] 0.8105314 0.6873338 0.9182667 0.3612247 0.8161310 0.6662936 0.3291508
#> [27,] 0.7972389 0.6746021 0.9308745 0.4725983 0.8362457 0.7307685 0.4882815
#> [28,] 0.8083320 0.7138510 0.9018874 0.3879945 0.7953908 0.6717780 0.3376180
#> [29,] 0.8354262 0.7069624 0.9379373 0.4592279 0.8564952 0.7397721 0.4286056
#> [30,] 0.6845968 0.5659083 0.8832084 0.2884743 0.7261921 0.5461776 0.2854126
#> [31,] 0.7909079 0.7077042 0.8994129 0.3590793 0.7660595 0.6420620 0.2910876
#> [32,] 0.7920599 0.6995452 0.9131832 0.4009697 0.7818242 0.6508972 0.3712136
#> [33,] 0.8507358 0.7266339 0.9515226 0.5364124 0.8750110 0.7907499 0.5309121
#> [34,] 0.6768606 0.5667472 0.8719889 0.3191104 0.7041236 0.5496529 0.2729443
#> [35,] 0.8491495 0.7200862 0.9450299 0.4614597 0.8709154 0.7472406 0.4229765
#> [36,] 0.8343050 0.7328085 0.9214661 0.4260016 0.8191661 0.6851511 0.3994601
#> [37,] 0.7888641 0.6906804 0.8992650 0.3897961 0.7796596 0.6516680 0.3272185
#> [38,] 0.8042806 0.6881019 0.9496789 0.3900250 0.8265329 0.7101841 0.3566321
#> [39,] 0.7102990 0.5860393 0.9184249 0.3115492 0.7496754 0.5910403 0.2841409
#> [40,] 0.7427931 0.6256323 0.9071150 0.3552015 0.7678116 0.6431801 0.2943143
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6050701 0.3908135 0.8885590
#>  [2,] 0.6927525 0.4135083 0.9117738
#>  [3,] 0.7234224 0.5002614 0.9050867
#>  [4,] 0.6919360 0.4699138 0.8957703
#>  [5,] 0.6226146 0.3883945 0.8927172
#>  [6,] 0.6495717 0.4217071 0.9018897
#>  [7,] 0.6968449 0.4686885 0.9145012
#>  [8,] 0.6025467 0.3777052 0.8634452
#>  [9,] 0.7620533 0.4832193 0.9289823
#> [10,] 0.6335076 0.4439340 0.8789314
#> [11,] 0.7273772 0.5147002 0.9151092
#> [12,] 0.5539112 0.3445706 0.8375415
#> [13,] 0.4579735 0.2323680 0.8097249
#> [14,] 0.5171528 0.3700087 0.8526027
#> [15,] 0.6592524 0.3428613 0.9233290
#> [16,] 0.6999629 0.4944334 0.9149793
#> [17,] 0.5726153 0.3843980 0.8900751
#> [18,] 0.6542380 0.4350540 0.9042708
#> [19,] 0.7167493 0.5150632 0.9084384
#> [20,] 0.5690700 0.3822223 0.8400392
#> [21,] 0.5802548 0.3843218 0.8694571
#> [22,] 0.7221997 0.5171302 0.9094447
#> [23,] 0.7021283 0.4815615 0.9111208
#> [24,] 0.7384576 0.5113257 0.9187667
#> [25,] 0.6957776 0.4610507 0.8932830
#> [26,] 0.5895862 0.4228181 0.8830257
#> [27,] 0.6872717 0.4369730 0.8896993
#> [28,] 0.6207026 0.4370951 0.8881676
#> [29,] 0.6906401 0.4817541 0.9059896
#> [30,] 0.4752632 0.3419772 0.8257696
#> [31,] 0.5886847 0.3749869 0.8870554
#> [32,] 0.6408333 0.4258961 0.8892078
#> [33,] 0.7589800 0.4983875 0.9226198
#> [34,] 0.5003901 0.3099072 0.8252570
#> [35,] 0.7038220 0.4892966 0.9173183
#> [36,] 0.6697344 0.4546220 0.9011383
#> [37,] 0.6211371 0.4130403 0.8863955
#> [38,] 0.6527887 0.3655901 0.9110523
#> [39,] 0.5365338 0.3106412 0.8613018
#> [40,] 0.5780559 0.3368069 0.8604009
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6721016 0.3278984
#>  [2,] 0.6785800 0.3214200
#>  [3,] 0.6809706 0.3190294
#>  [4,] 0.6759667 0.3240333
#>  [5,] 0.6732491 0.3267509
#>  [6,] 0.6785592 0.3214408
#>  [7,] 0.6818497 0.3181503
#>  [8,] 0.6684557 0.3315443
#>  [9,] 0.6870075 0.3129925
#> [10,] 0.6720731 0.3279269
#> [11,] 0.6844270 0.3155730
#> [12,] 0.6592753 0.3407247
#> [13,] 0.6388174 0.3611826
#> [14,] 0.6616152 0.3383848
#> [15,] 0.6785972 0.3214028
#> [16,] 0.6840488 0.3159512
#> [17,] 0.6685563 0.3314437
#> [18,] 0.6771454 0.3228546
#> [19,] 0.6849707 0.3150293
#> [20,] 0.6620087 0.3379913
#> [21,] 0.6683401 0.3316599
#> [22,] 0.6832065 0.3167935
#> [23,] 0.6823311 0.3176689
#> [24,] 0.6876432 0.3123568
#> [25,] 0.6765832 0.3234168
#> [26,] 0.6734676 0.3265324
#> [27,] 0.6759205 0.3240795
#> [28,] 0.6745874 0.3254126
#> [29,] 0.6818375 0.3181625
#> [30,] 0.6527130 0.3472870
#> [31,] 0.6709937 0.3290063
#> [32,] 0.6723822 0.3276178
#> [33,] 0.6862146 0.3137854
#> [34,] 0.6525454 0.3474546
#> [35,] 0.6844910 0.3155090
#> [36,] 0.6782047 0.3217953
#> [37,] 0.6720939 0.3279061
#> [38,] 0.6769007 0.3230993
#> [39,] 0.6596797 0.3403203
#> [40,] 0.6655206 0.3344794
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]     [,3]         [,4]      [,5]       [,6]
#>  [1,] 1.3515205 0.8248000 2.280054 -0.599290846 1.2864580  0.6792906
#>  [2,] 1.4822166 0.8712975 2.650216 -0.196920195 1.6332162  0.9563670
#>  [3,] 1.5582266 0.9330528 2.596575  0.003436534 1.6601262  1.0209558
#>  [4,] 1.4381794 0.8536118 2.291085 -0.237935624 1.4804573  0.8009955
#>  [5,] 1.4548284 0.8632665 2.258371 -0.494195534 1.3360241  0.7384588
#>  [6,] 1.5796593 0.8535927 2.692916 -0.384373507 1.6101295  0.8722216
#>  [7,] 1.6318504 0.9435036 2.634262 -0.186266844 1.6390798  0.9528516
#>  [8,] 1.1837314 0.5766040 2.380243 -0.476433701 1.3312750  0.6492439
#>  [9,] 1.7737645 1.0056019 3.103622  0.103707817 2.0117511  1.3462041
#> [10,] 1.2367171 0.6840455 2.346397 -0.242257498 1.4469683  0.7468802
#> [11,] 1.6334468 0.9161675 2.810044  0.118859310 1.8512198  1.2076792
#> [12,] 1.0218376 0.5105809 2.151756 -0.682552497 1.1235505  0.3881100
#> [13,] 0.4464646 0.2844770 1.588760 -1.242822083 0.3690476 -0.3312657
#> [14,] 1.1025457 0.6417927 1.915613 -0.789522447 1.0065528  0.3559440
#> [15,] 1.5324816 0.8425236 3.108247 -0.541738846 1.5899617  0.8575063
#> [16,] 1.7155732 0.9577041 2.749788 -0.118475938 1.8399937  1.1039455
#> [17,] 1.1928485 0.6099149 2.437705 -0.630694490 1.3144931  0.4691227
#> [18,] 1.4092732 0.7931514 2.501792 -0.234580553 1.5114380  0.9647140
#> [19,] 1.8109451 1.0488957 2.636829 -0.059874437 1.8420228  1.1825282
#> [20,] 1.0798433 0.5631288 1.962023 -0.763920960 1.2114436  0.4514901
#> [21,] 1.3129375 0.7235531 2.261853 -0.709386931 1.3022874  0.4886655
#> [22,] 1.6132486 0.8591803 2.828015  0.102279403 1.8304245  1.2018165
#> [23,] 1.6307707 0.8589854 2.776773 -0.008497114 1.7610498  1.0968123
#> [24,] 1.9077080 1.0866296 2.833133 -0.081521149 1.9868517  1.2547842
#> [25,] 1.3822744 0.7507361 2.651003 -0.055388940 1.6141153  1.0328774
#> [26,] 1.4534665 0.7876837 2.419026 -0.570052543 1.4903516  0.6914689
#> [27,] 1.3691262 0.7290741 2.600201 -0.109716887 1.6305553  0.9985252
#> [28,] 1.4392081 0.9141616 2.218373 -0.455750293 1.3577319  0.7162377
#> [29,] 1.6245827 0.8806764 2.715538 -0.163451415 1.7864800  1.0447842
#> [30,] 0.7749781 0.2651762 2.023170 -0.902805558 0.9753879  0.1852384
#> [31,] 1.3304065 0.8842599 2.190719 -0.579362602 1.1861932  0.5843252
#> [32,] 1.3373868 0.8451331 2.353136 -0.401426390 1.2763288  0.6229852
#> [33,] 1.7403840 0.9776110 2.976966  0.145907736 1.9460106  1.3294513
#> [34,] 0.7393814 0.2685919 1.918660 -0.757863234 0.8670120  0.1992685
#> [35,] 1.7279462 0.9448891 2.844427 -0.154467656 1.9090764  1.0839491
#> [36,] 1.6164500 1.0089188 2.462435 -0.298183504 1.5107078  0.7775465
#> [37,] 1.3180920 0.8033021 2.189085 -0.448169506 1.2636839  0.6263793
#> [38,] 1.4132661 0.7912605 2.937700 -0.447206981 1.5612519  0.8962782
#> [39,] 0.8968366 0.3476159 2.421136 -0.792886399 1.0968819  0.3682675
#> [40,] 1.0605368 0.5135242 2.278907 -0.596252577 1.1959955  0.5891937
#>              [,7]         [,8]         [,9]    [,10]
#>  [1,] -0.80015009  0.426636231 -0.443893904 2.076105
#>  [2,] -0.23735728  0.813019148 -0.349480923 2.335488
#>  [3,] -0.03042942  0.961501785  0.001045542 2.255067
#>  [4,] -0.24210661  0.809185712 -0.120490207 2.151087
#>  [5,] -0.61931815  0.500660634 -0.454065852 2.118801
#>  [6,] -0.59296018  0.617157199 -0.315769305 2.218400
#>  [7,] -0.40958624  0.832318356 -0.125410279 2.369877
#>  [8,] -0.65295889  0.416087752 -0.499299998 1.844204
#>  [9,]  0.08785293  1.163970055 -0.067147932 2.571161
#> [10,] -0.31127447  0.547294393 -0.225210935 1.982350
#> [11,]  0.01717110  0.981355869  0.058817734 2.377678
#> [12,] -0.56424639  0.216486301 -0.642991618 1.640049
#> [13,] -1.44948850 -0.168503772 -1.194988211 1.448224
#> [14,] -0.98781955  0.068638059 -0.532179567 1.755162
#> [15,] -0.77050970  0.659964645 -0.650568947 2.488462
#> [16,] -0.26519559  0.847121399 -0.022267218 2.376007
#> [17,] -0.82411347  0.292529543 -0.470922217 2.091508
#> [18,] -0.36693996  0.637720523 -0.261260149 2.245605
#> [19,] -0.15392015  0.928393691  0.060271098 2.294716
#> [20,] -0.69104602  0.278057936 -0.480126369 1.658520
#> [21,] -0.84039231  0.323819435 -0.471244398 1.896167
#> [22,]  0.06043940  0.955398995  0.068547484 2.306874
#> [23,] -0.14784585  0.857453466 -0.073787639 2.327398
#> [24,] -0.21115565  1.037967235  0.045310372 2.425707
#> [25,]  0.05168589  0.827270837 -0.156113559 2.124723
#> [26,] -0.71202826  0.362254906 -0.311215359 2.021400
#> [27,] -0.04688253  0.787394876 -0.253456255 2.087674
#> [28,] -0.67392731  0.492531375 -0.252960094 2.072160
#> [29,] -0.28754259  0.803113687 -0.073015944 2.265623
#> [30,] -0.91776949 -0.099027982 -0.654495611 1.555937
#> [31,] -0.89010773  0.358530473 -0.510881317 2.061010
#> [32,] -0.52701391  0.578982551 -0.298615045 2.082675
#> [33,]  0.12380629  1.147095590 -0.006449858 2.478487
#> [34,] -0.97973526  0.001560313 -0.800553366 1.552379
#> [35,] -0.31056625  0.865565020 -0.042820067 2.406456
#> [36,] -0.40771529  0.706984106 -0.182012845 2.209936
#> [37,] -0.72079258  0.494377523 -0.351411156 2.054440
#> [38,] -0.59001191  0.631320016 -0.551182603 2.326552
#> [39,] -0.92401324  0.146395991 -0.797123475 1.826144
#> [40,] -0.87452195  0.314797939 -0.677556271 1.818624
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.4095686 -0.3081374
#>  [2,] 0.4185579 -0.3286956
#>  [3,] 0.4192085 -0.3390272
#>  [4,] 0.4112462 -0.3240513
#>  [5,] 0.4110548 -0.3118624
#>  [6,] 0.4223533 -0.3248050
#>  [7,] 0.4284985 -0.3337868
#>  [8,] 0.3942920 -0.3069168
#>  [9,] 0.4338892 -0.3522770
#> [10,] 0.3987249 -0.3188514
#> [11,] 0.4254860 -0.3487059
#> [12,] 0.3714040 -0.2886622
#> [13,] 0.3339564 -0.2362788
#> [14,] 0.3837694 -0.2867313
#> [15,] 0.4278017 -0.3195307
#> [16,] 0.4307467 -0.3416948
#> [17,] 0.3993132 -0.3023495
#> [18,] 0.4133772 -0.3273067
#> [19,] 0.4317143 -0.3449960
#> [20,] 0.3808006 -0.2914580
#> [21,] 0.4006743 -0.3000130
#> [22,] 0.4206070 -0.3479400
#> [23,] 0.4230357 -0.3414697
#> [24,] 0.4404994 -0.3486247
#> [25,] 0.4043975 -0.3337159
#> [26,] 0.4103809 -0.3135298
#> [27,] 0.4049396 -0.3301470
#> [28,] 0.4141604 -0.3148471
#> [29,] 0.4247396 -0.3374897
#> [30,] 0.3556737 -0.2753125
#> [31,] 0.4091967 -0.3034861
#> [32,] 0.4065126 -0.3124666
#> [33,] 0.4301179 -0.3523633
#> [34,] 0.3566316 -0.2736151
#> [35,] 0.4330596 -0.3414287
#> [36,] 0.4212698 -0.3242637
#> [37,] 0.4080855 -0.3095852
#> [38,] 0.4186441 -0.3209210
#> [39,] 0.3749877 -0.2868794
#> [40,] 0.3897092 -0.2982853
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3278984 -0.3278984
#>  [2,] -0.6785800  0.6785800
#>  [3,] -0.6809706  0.6809706
#>  [4,] -0.6759667  0.6759667
#>  [5,] -0.6732491  0.6732491
#>  [6,]  0.3214408 -0.3214408
#>  [7,]  0.3181503 -0.3181503
#>  [8,] -0.6684557  0.6684557
#>  [9,] -0.6870075  0.6870075
#> [10,]  0.3279269 -0.3279269
#> [11,]  0.3155730 -0.3155730
#> [12,] -0.6592753  0.6592753
#> [13,]  0.3611826 -0.3611826
#> [14,]  0.3383848 -0.3383848
#> [15,] -0.6785972  0.6785972
#> [16,]  0.3159512 -0.3159512
#> [17,]  0.3314437 -0.3314437
#> [18,] -0.6771454  0.6771454
#> [19,] -0.6849707  0.6849707
#> [20,] -0.6620087  0.6620087
#> [21,]  0.3316599 -0.3316599
#> [22,] -0.6832065  0.6832065
#> [23,]  0.3176689 -0.3176689
#> [24,] -0.6876432  0.6876432
#> [25,] -0.6765832  0.6765832
#> [26,] -0.6734676  0.6734676
#> [27,] -0.6759205  0.6759205
#> [28,]  0.3254126 -0.3254126
#> [29,]  0.3181625 -0.3181625
#> [30,]  0.3472870 -0.3472870
#> [31,]  0.3290063 -0.3290063
#> [32,] -0.6723822  0.6723822
#> [33,] -0.6862146  0.6862146
#> [34,]  0.3474546 -0.3474546
#> [35,]  0.3155090 -0.3155090
#> [36,] -0.6782047  0.6782047
#> [37,]  0.3279061 -0.3279061
#> [38,] -0.6769007  0.6769007
#> [39,]  0.3403203 -0.3403203
#> [40,] -0.6655206  0.6655206
#> 
#> $L
#> [1] 0.6990784 0.7316453 0.7761156 0.7038885 0.6991226 0.7821817
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319