Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13          V14
#>  [1,] -0.2454804 -0.05681907 -0.174806253 -0.100596600 -0.02626185 -0.022982969
#>  [2,] -0.3190683 -0.02874845 -0.048218406 -0.082594615 -0.03462383  0.008511457
#>  [3,] -0.1588923 -0.18928365 -0.143425647 -0.096624865 -0.09935721 -0.174812666
#>  [4,] -0.3330493 -0.09214952 -0.094944389 -0.009027574 -0.03116862 -0.060452645
#>  [5,] -0.3515567  0.02351509 -0.005902298 -0.043635088 -0.10738299  0.055901963
#>  [6,] -0.3814902 -0.21219350 -0.171656387 -0.065932762 -0.02773329 -0.056478246
#>  [7,] -0.2290241 -0.13436051 -0.091615613 -0.151394826 -0.15472427 -0.122054904
#>  [8,] -0.2708659 -0.26379582 -0.211038460 -0.142935162 -0.09496610 -0.156115781
#>  [9,] -0.2605264 -0.14384008 -0.084429077 -0.167699997 -0.07303858 -0.132844281
#> [10,] -0.3873187 -0.13789137 -0.080085102 -0.038978196 -0.15572936 -0.086271926
#>                V15          V16          V17          V18         V19
#>  [1,] -0.100973709 -0.046234376 -0.009229436  0.036138239  0.04886040
#>  [2,]  0.086397303 -0.006768567  0.118121696  0.009334706  0.14038326
#>  [3,]  0.001112597 -0.056321842  0.009837626  0.044942792  0.03442477
#>  [4,]  0.058171219 -0.038994521  0.006910125  0.109058483  0.12771410
#>  [5,]  0.086142914  0.114380985  0.116770754  0.185108595  0.11596483
#>  [6,] -0.129467946 -0.068565595  0.102190138 -0.011532517  0.02239264
#>  [7,] -0.014530371  0.072581313 -0.042398981 -0.010506097  0.07368146
#>  [8,] -0.118335210 -0.040799633 -0.117420123 -0.167332135 -0.15431251
#>  [9,] -0.109323163  0.010953512 -0.087751268 -0.013515202 -0.06538516
#> [10,] -0.137361050  0.001882114 -0.071980830  0.059295640 -0.02606848
#>               V2         V20         V21         V22          V23           V24
#>  [1,] -0.2344531 -0.04880720  0.07682457 -0.05104277 -0.002277207 -0.0093349836
#>  [2,] -0.3329762  0.08583125  0.10605213  0.02336585  0.042111338  0.1512994771
#>  [3,] -0.2889791 -0.09492338  0.11626980  0.03730708  0.076178422  0.0007759186
#>  [4,] -0.3213493  0.07492352  0.13120910  0.09560164  0.056803080  0.2017964111
#>  [5,] -0.3216808  0.14347971  0.06831618  0.14348143  0.208128579  0.1736289742
#>  [6,] -0.3735031  0.01976707  0.03641697  0.12630421  0.085192180 -0.0061854123
#>  [7,] -0.2011445  0.06334422 -0.03176551  0.05755427  0.058896050  0.0216055306
#>  [8,] -0.2876574 -0.00931711 -0.10734213  0.01589120  0.019594539  0.0673812237
#>  [9,] -0.2754283  0.03294801  0.02280221  0.07630289  0.037908420  0.1886682160
#> [10,] -0.3182405  0.07288321 -0.03175608  0.14712781  0.094641157  0.0253502601
#>                V25           V26        V27          V28         V29         V3
#>  [1,]  0.057069909  0.0005507647 0.11868523 -0.040385586  0.09222448 -0.3041756
#>  [2,]  0.151774429  0.1182762699 0.13359094  0.269891101  0.18002993 -0.2074930
#>  [3,]  0.147724723  0.0857375980 0.03260048  0.032343278  0.02522860 -0.3354495
#>  [4,]  0.122869189  0.2608299754 0.13288029  0.154016598  0.13896348 -0.1923701
#>  [5,]  0.221915162  0.2554726952 0.12265467  0.093976397  0.10789062 -0.3012173
#>  [6,]  0.172464391  0.1787668175 0.19343190  0.214604412  0.13398258 -0.4219057
#>  [7,] -0.082498081  0.1324622380 0.13473269 -0.003975496 -0.04963006 -0.3486849
#>  [8,]  0.007021027 -0.0305742138 0.11776494 -0.018418764 -0.06864791 -0.3014836
#>  [9,]  0.118388713  0.1842443218 0.14962015  0.289103521  0.23238043 -0.1621415
#> [10,]  0.061810296  0.0767698152 0.07618233  0.105768277  0.14590369 -0.3544003
#>                V30          V31          V32           V33         V34
#>  [1,] -0.001854185  0.005589461 -0.067912415 -0.1115924883 -0.05601027
#>  [2,]  0.207629579  0.245509561  0.120692979  0.1040739698  0.14078298
#>  [3,]  0.012028305 -0.029622671 -0.039006333 -0.0418206167 -0.07861799
#>  [4,]  0.084191514  0.013393048 -0.068619030  0.0779628493  0.12099600
#>  [5,]  0.111841329  0.034794456  0.155625632  0.1423784470  0.03166421
#>  [6,]  0.124726137  0.055993131  0.070733318  0.0337844436  0.07171644
#>  [7,] -0.061431690  0.001134241 -0.091788367 -0.1683751041 -0.19479177
#>  [8,]  0.005164068 -0.015321609 -0.058862415 -0.1377605015 -0.09558362
#>  [9,]  0.133494579  0.067785621  0.006929199  0.0005694986  0.02357274
#> [10,]  0.094752538  0.031965592  0.021507108 -0.0809977077 -0.07240951
#>               V35          V36         V37          V38         V39         V4
#>  [1,] -0.09158190 -0.001927076  0.03735554  0.002958163 -0.13338745 -0.1384834
#>  [2,]  0.15646951  0.128266469 -0.06485832 -0.052149275  0.07815242 -0.2140630
#>  [3,] -0.17252167 -0.013906411 -0.04726941 -0.118249775 -0.01325015 -0.1876131
#>  [4,]  0.10240546 -0.102691356 -0.08281879  0.023984725 -0.03027966 -0.3240998
#>  [5,]  0.12896990  0.014806148  0.03211718 -0.001458603 -0.02483437 -0.3448921
#>  [6,] -0.09895475  0.014082050  0.01884410 -0.014408338 -0.16342583 -0.4061315
#>  [7,] -0.20192170 -0.193168927 -0.15789208 -0.039671522 -0.03139238 -0.2635334
#>  [8,] -0.06021987 -0.096790707 -0.08839475 -0.093831638 -0.11311947 -0.2962455
#>  [9,]  0.13918471 -0.026230327 -0.03379238  0.059439853  0.05409243 -0.2330478
#> [10,] -0.05917078 -0.102968544 -0.03667473 -0.139125253 -0.03618926 -0.2737274
#>               V40          V41         V42         V43         V44         V45
#>  [1,]  0.01469931 -0.113101687 -0.13401318 -0.11544734 -0.09403960 -0.18879592
#>  [2,] -0.03168932  0.052028269 -0.08150417  0.02550937  0.01689766 -0.05628414
#>  [3,] -0.09485417 -0.195612352 -0.20287855 -0.09959736 -0.21739087 -0.10308462
#>  [4,] -0.05945109  0.015423009 -0.06392367 -0.14128995 -0.20446318 -0.20036868
#>  [5,] -0.05269698 -0.006631553  0.02327752 -0.16474613 -0.16060410 -0.08939199
#>  [6,] -0.07466940 -0.223586984 -0.07477685 -0.06489982 -0.08541232 -0.08916977
#>  [7,] -0.10886147 -0.183689541 -0.11214063 -0.21627749 -0.10624009 -0.20424087
#>  [8,] -0.20015920 -0.155853565 -0.12480386 -0.10517180 -0.19924502 -0.19558689
#>  [9,] -0.01695695 -0.066298253 -0.01389775 -0.04550064 -0.14206358 -0.14802979
#> [10,] -0.05209888 -0.068324230 -0.11862303 -0.05809306 -0.16616937 -0.16041944
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.07549571 -0.1864069 -0.1474401 -0.2583541 -0.2410743 -0.1915088
#>  [2,] -0.12864480 -0.1048900 -0.1928885 -0.3016916 -0.2480193 -0.3860650
#>  [3,] -0.21342204 -0.2341534 -0.1270224 -0.3063199 -0.1574250 -0.3713869
#>  [4,] -0.09607103 -0.2279577 -0.2242261 -0.1577469 -0.2612005 -0.3981401
#>  [5,] -0.10595514 -0.1604166 -0.2136232 -0.2205429 -0.3074057 -0.2445081
#>  [6,] -0.12172690 -0.2849501 -0.2154779 -0.3787596 -0.2972281 -0.3625670
#>  [7,] -0.25458069 -0.2473451 -0.1510875 -0.2373313 -0.1626538 -0.1802295
#>  [8,] -0.09117801 -0.1634297 -0.2947754 -0.3339193 -0.1728895 -0.3624815
#>  [9,] -0.17542950 -0.2151067 -0.1488197 -0.3288863 -0.1828187 -0.3312623
#> [10,] -0.25184623 -0.1164326 -0.2790481 -0.2367764 -0.2108177 -0.4791656
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.1919028 -0.2178826 -0.2841062 -0.2467139 -0.2479130 -0.2383593
#>  [2,] -0.4043287 -0.3809444 -0.4026744 -0.4196184 -0.3989123 -0.3758840
#>  [3,] -0.3835820 -0.3452442 -0.3729710 -0.2052124 -0.3734459 -0.2854331
#>  [4,] -0.2182333 -0.3316261 -0.3529998 -0.2066919 -0.3142050 -0.3781139
#>  [5,] -0.3047021 -0.4310419 -0.4140690 -0.4149699 -0.3880172 -0.3458697
#>  [6,] -0.3248431 -0.3315644 -0.3984078 -0.4384329 -0.3690506 -0.4277367
#>  [7,] -0.3179430 -0.3049238 -0.2444450 -0.2913930 -0.3728754 -0.1808809
#>  [8,] -0.2942827 -0.3356279 -0.2284697 -0.3149750 -0.3030894 -0.2185195
#>  [9,] -0.2949751 -0.2567578 -0.3468167 -0.2968887 -0.2575127 -0.3434827
#> [10,] -0.4682651 -0.4291672 -0.3005512 -0.4178880 -0.4691562 -0.4438364
#>              V57        V58        V59          V6        V60          V7
#>  [1,] -0.2705485 -0.2895243 -0.2841127 -0.17939737 -0.2798588 -0.16406112
#>  [2,] -0.4290137 -0.3993719 -0.3519702 -0.19585553 -0.3818379 -0.12317216
#>  [3,] -0.2437467 -0.3357520 -0.2363528 -0.23060772 -0.2911013 -0.14438150
#>  [4,] -0.4097452 -0.3499233 -0.2514972 -0.15173351 -0.2319228 -0.11525733
#>  [5,] -0.4228076 -0.4239939 -0.4526713 -0.24847195 -0.3114379 -0.08475715
#>  [6,] -0.4004329 -0.3737858 -0.4834005 -0.24319820 -0.4206401 -0.30856407
#>  [7,] -0.3280511 -0.3022394 -0.3511740 -0.27032587 -0.2229204 -0.15599395
#>  [8,] -0.2279832 -0.3922462 -0.3761515 -0.22154483 -0.3832694 -0.15058773
#>  [9,] -0.3410523 -0.2858100 -0.2564843 -0.06544633 -0.2651341 -0.13121698
#> [10,] -0.4151355 -0.3627120 -0.3538915 -0.24769832 -0.3572810 -0.25710090
#>               V8           V9
#>  [1,] -0.1863434 -0.111350190
#>  [2,] -0.1905120 -0.181688093
#>  [3,] -0.2110696 -0.205471993
#>  [4,] -0.1961847 -0.007779695
#>  [5,] -0.1434299 -0.071583517
#>  [6,] -0.2162927 -0.214418913
#>  [7,] -0.2047844 -0.132272868
#>  [8,] -0.1620181 -0.146356048
#>  [9,] -0.1674982 -0.148398319
#> [10,] -0.2055530 -0.158233419
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] -0.04078230  0.00651411 -0.07372510 0.15364751  0.05553862  0.06281734
#> [2,] -0.06110868 -0.02525836  0.08430783 0.02966829 -0.13512049 -0.02975531
#>             [,7]         [,8]        [,9]       [,10]
#> [1,] -0.07234521 -0.059896763  0.03666674 -0.02915439
#> [2,] -0.03106456  0.009266126 -0.12976707 -0.04611278
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  1.807052e-05  9.794461e-05  0.0001339247  0.0001561813  0.0001522614
#>  [2,]  9.757075e-05  5.816258e-04  0.0006768858  0.0006525800  0.0006542130
#>  [3,]  2.302883e-04  1.704082e-03  0.0019375841  0.0018091886  0.0017939082
#>  [4,] -3.768672e-05 -3.353980e-04 -0.0002757523 -0.0002050810 -0.0002546732
#>  [5,] -5.625788e-05 -4.390681e-04 -0.0003981662 -0.0003289766 -0.0003826874
#>  [6,] -3.910495e-05 -3.496027e-04 -0.0003098840 -0.0002325402 -0.0002719939
#>  [7,]  6.665459e-05  4.768378e-04  0.0005674655  0.0005574817  0.0005459721
#>  [8,]  5.417021e-05  4.057587e-04  0.0004665072  0.0004431743  0.0004359323
#>  [9,] -1.038667e-04 -7.859538e-04 -0.0007571509 -0.0006452708 -0.0007170381
#> [10,]  5.806080e-05  3.683100e-04  0.0004531472  0.0004537185  0.0004399637
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0001433409  0.0001585090  1.973120e-04  2.136472e-04  2.265815e-04
#>  [2,]  0.0006101371  0.0004955983  4.018378e-04  3.907981e-04  4.117722e-04
#>  [3,]  0.0015093751  0.0008632835  2.975341e-04  3.411370e-04  5.851055e-04
#>  [4,] -0.0002904027 -0.0001586548  1.350127e-06  3.891689e-05  2.020913e-05
#>  [5,] -0.0004175442 -0.0002933374 -1.486667e-04 -1.185944e-04 -1.226100e-04
#>  [6,] -0.0002777214 -0.0001129463  6.671616e-05  9.665874e-05  5.960850e-05
#>  [7,]  0.0004525026  0.0002890877  1.582827e-04  1.937298e-04  2.665845e-04
#>  [8,]  0.0003492050  0.0001871932  4.141903e-05  6.564333e-05  1.234651e-04
#>  [9,] -0.0007313454 -0.0004559255 -1.700569e-04 -1.147941e-04 -1.735765e-04
#> [10,]  0.0003821132  0.0003091256  2.830270e-04  3.041749e-04  3.414860e-04
#>                 V19            V2           V20           V21           V22
#>  [1,]  2.776310e-04  2.069624e-05  3.157962e-04  0.0003365531  0.0003519207
#>  [2,]  6.367905e-04  1.097614e-04  8.207037e-04  0.0010264402  0.0010981440
#>  [3,]  1.107672e-03  2.677812e-04  1.433165e-03  0.0021318562  0.0025037275
#>  [4,] -2.706407e-05 -4.774645e-05 -8.420957e-05 -0.0002193529 -0.0001887168
#>  [5,] -2.041445e-04 -6.933495e-05 -3.059347e-04 -0.0004681135 -0.0004247406
#>  [6,] -2.583003e-06 -4.794870e-05 -6.315195e-05 -0.0002179706 -0.0002232828
#>  [7,]  4.161437e-04  7.828060e-05  5.093882e-04  0.0007080940  0.0008180753
#>  [8,]  2.086505e-04  6.599951e-05  2.510068e-04  0.0004311129  0.0005449131
#>  [9,] -3.786981e-04 -1.234622e-04 -5.839409e-04 -0.0008982511 -0.0008827223
#> [10,]  5.006940e-04  6.563870e-05  6.123080e-04  0.0007294544  0.0008023666
#>                 V23           V24           V25           V26           V27
#>  [1,]  3.885374e-04  4.143510e-04  3.948173e-04  4.078322e-04  4.152506e-04
#>  [2,]  9.595965e-04  8.497722e-04  8.736378e-04  9.677574e-04  9.815942e-04
#>  [3,]  1.922866e-03  1.449793e-03  1.610411e-03  1.820231e-03  1.897415e-03
#>  [4,]  6.003559e-06  1.268681e-04  1.952558e-05 -6.431503e-05 -7.606823e-05
#>  [5,] -2.138212e-04 -9.609055e-05 -2.219946e-04 -3.277739e-04 -3.308955e-04
#>  [6,] -1.703035e-05  1.266048e-04  2.497669e-05 -5.281682e-05 -6.147791e-05
#>  [7,]  6.732227e-04  5.536294e-04  5.745201e-04  6.216286e-04  6.560772e-04
#>  [8,]  3.725931e-04  2.483233e-04  3.066644e-04  3.441711e-04  3.678616e-04
#>  [9,] -5.468559e-04 -2.955653e-04 -4.271937e-04 -6.020878e-04 -6.441361e-04
#> [10,]  7.743332e-04  7.460690e-04  7.447911e-04  7.954026e-04  8.070695e-04
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0004021329  0.0003710700  2.207088e-05  0.0003046069  0.0002384457
#>  [2,]  0.0010960574  0.0009945678  1.079832e-04  0.0009311738  0.0006332211
#>  [3,]  0.0022323237  0.0018257135  2.722366e-04  0.0017473380  0.0010732974
#>  [4,] -0.0002119939 -0.0001737952 -4.950590e-05 -0.0003087832 -0.0002438667
#>  [5,] -0.0004724633 -0.0004203052 -7.280262e-05 -0.0005393628 -0.0004210773
#>  [6,] -0.0002013259 -0.0001556339 -4.827665e-05 -0.0002772530 -0.0001981846
#>  [7,]  0.0007404091  0.0006212960  7.869719e-05  0.0005492754  0.0003383253
#>  [8,]  0.0004493536  0.0003518129  6.689739e-05  0.0003572942  0.0002227779
#>  [9,] -0.0009220149 -0.0008087122 -1.220277e-04 -0.0009434988 -0.0006705525
#> [10,]  0.0008368439  0.0007499244  6.683131e-05  0.0006522687  0.0004535161
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0001987468  0.0001812873  1.627504e-04  1.373175e-04  1.331647e-04
#>  [2,]  0.0006921586  0.0005801713  2.901186e-04  1.554113e-04  1.221733e-04
#>  [3,]  0.0013774925  0.0009607733 -1.160602e-04 -5.467459e-04 -5.911979e-04
#>  [4,] -0.0003099747 -0.0002417090 -9.232572e-05 -6.989691e-05 -2.575174e-05
#>  [5,] -0.0004717167 -0.0003699208 -2.013658e-04 -1.669278e-04 -1.172300e-04
#>  [6,] -0.0002898954 -0.0002197612 -2.358120e-05  2.272748e-05  6.238868e-05
#>  [7,]  0.0004120881  0.0002919934 -4.841092e-06 -1.309234e-04 -1.357361e-04
#>  [8,]  0.0003063698  0.0001784180 -1.137448e-04 -2.154128e-04 -2.150311e-04
#>  [9,] -0.0008027019 -0.0006746447 -3.067828e-04 -1.926773e-04 -1.065146e-04
#> [10,]  0.0004490878  0.0003654197  1.860870e-04  8.169139e-05  6.610576e-05
#>                 V37           V38           V39            V4           V40
#>  [1,]  1.426663e-04  0.0001497928  0.0001488872  2.951160e-05  0.0001259981
#>  [2,]  2.554713e-04  0.0004888982  0.0005030727  1.780462e-04  0.0003636154
#>  [3,] -9.819730e-05  0.0007614505  0.0009846500  4.715896e-04  0.0008111431
#>  [4,] -5.274930e-05 -0.0001978384 -0.0002092999 -8.145191e-05 -0.0001898930
#>  [5,] -1.642360e-04 -0.0003243142 -0.0003143185 -1.144280e-04 -0.0002659282
#>  [6,]  2.282561e-05 -0.0001511763 -0.0001874177 -8.627692e-05 -0.0001692766
#>  [7,]  1.241991e-05  0.0002376026  0.0002922250  1.362934e-04  0.0002324330
#>  [8,] -7.030712e-05  0.0001303255  0.0001689371  1.138064e-04  0.0001471514
#>  [9,] -1.821848e-04 -0.0005352051 -0.0005928257 -2.100733e-04 -0.0004874803
#> [10,]  1.608713e-04  0.0003089131  0.0003370956  1.063329e-04  0.0002654022
#>                 V41           V42           V43           V44           V45
#>  [1,]  0.0001082244  0.0001134000  0.0001042115  0.0001003752  8.639355e-05
#>  [2,]  0.0005792228  0.0007659117  0.0007750732  0.0006468901  7.735491e-04
#>  [3,]  0.0016809979  0.0021432850  0.0022341628  0.0018141014  2.169060e-03
#>  [4,] -0.0004328059 -0.0004983814 -0.0004977345 -0.0003723453 -4.777083e-04
#>  [5,] -0.0005310906 -0.0006146724 -0.0006121250 -0.0004870963 -6.087818e-04
#>  [6,] -0.0004275813 -0.0005124122 -0.0005206053 -0.0003905811 -5.029140e-04
#>  [7,]  0.0004451208  0.0005801491  0.0006062243  0.0004921936  5.710527e-04
#>  [8,]  0.0003802257  0.0004843885  0.0005121801  0.0004179654  5.019689e-04
#>  [9,] -0.0009422520 -0.0011394395 -0.0011359293 -0.0008665597 -1.076517e-03
#> [10,]  0.0003743364  0.0004624382  0.0004640760  0.0004110631  4.641994e-04
#>                 V46           V47           V48           V49            V5
#>  [1,]  6.899493e-05  4.771429e-05  0.0000286216  0.0000152121  4.964248e-05
#>  [2,]  6.595716e-04  4.394384e-04  0.0003576057  0.0001982347  1.960755e-04
#>  [3,]  1.809399e-03  1.281146e-03  0.0010758373  0.0006339265  4.623225e-04
#>  [4,] -4.259582e-04 -3.028729e-04 -0.0002646246 -0.0001554468 -5.026101e-05
#>  [5,] -5.475387e-04 -3.835661e-04 -0.0003237906 -0.0001845631 -9.049649e-05
#>  [6,] -4.389203e-04 -3.099234e-04 -0.0002746528 -0.0001621081 -5.302615e-05
#>  [7,]  4.698660e-04  3.365295e-04  0.0002819209  0.0001666814  1.445385e-04
#>  [8,]  4.233957e-04  3.113367e-04  0.0002606041  0.0001535402  1.056487e-04
#>  [9,] -9.328099e-04 -6.425862e-04 -0.0005547178 -0.0003226738 -1.754158e-04
#> [10,]  3.885397e-04  2.624392e-04  0.0002011101  0.0001130945  1.320419e-04
#>                 V50           V51           V52           V53           V54
#>  [1,]  4.878845e-06  7.361899e-06  6.386480e-06  5.196700e-06  6.065336e-06
#>  [2,]  5.456753e-05  5.923738e-05  4.635404e-05  2.374652e-05  3.199577e-05
#>  [3,]  1.563338e-04  1.496209e-04  1.074998e-04  4.588227e-05  7.860085e-05
#>  [4,] -4.667701e-05 -3.456793e-05 -2.156697e-05 -9.275203e-06 -1.504345e-05
#>  [5,] -5.489044e-05 -4.732284e-05 -3.019901e-05 -1.398593e-05 -2.064756e-05
#>  [6,] -4.641271e-05 -3.482555e-05 -2.279127e-05 -8.802452e-06 -1.537917e-05
#>  [7,]  4.025006e-05  4.174570e-05  2.973078e-05  1.387803e-05  2.310282e-05
#>  [8,]  3.509861e-05  3.509853e-05  2.365061e-05  8.390481e-06  1.783196e-05
#>  [9,] -9.470402e-05 -7.957872e-05 -5.621370e-05 -2.691587e-05 -3.962779e-05
#> [10,]  2.947909e-05  3.354874e-05  2.643068e-05  1.412088e-05  1.896259e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  4.626669e-06  4.278666e-06  3.464976e-06  4.146518e-06  3.939688e-06
#>  [2,]  1.400651e-05  2.140762e-05  1.014091e-05  2.316162e-05  2.144187e-05
#>  [3,]  2.077389e-05  5.083493e-05  1.152386e-05  5.420446e-05  4.678386e-05
#>  [4,] -3.144659e-06 -9.872359e-06 -4.708290e-06 -9.085137e-06 -9.069245e-06
#>  [5,] -6.286391e-06 -1.411728e-05 -8.304519e-06 -1.309230e-05 -1.322966e-05
#>  [6,] -2.328549e-06 -9.566051e-06 -3.139250e-06 -9.932469e-06 -9.234706e-06
#>  [7,]  7.093139e-06  1.487941e-05  3.863652e-06  1.599402e-05  1.356871e-05
#>  [8,]  3.099529e-06  1.180213e-05  2.134744e-06  1.181877e-05  1.008843e-05
#>  [9,] -1.248975e-05 -2.481624e-05 -1.134805e-05 -2.636052e-05 -2.450502e-05
#> [10,]  9.270362e-06  1.279392e-05  5.795025e-06  1.374509e-05  1.208057e-05
#>                  V6           V60            V7            V8            V9
#>  [1,]  6.227143e-05  2.776618e-06  6.127191e-05  6.982082e-05  9.029572e-05
#>  [2,]  1.772390e-04  1.516154e-05  1.937722e-04  3.328314e-04  5.395396e-04
#>  [3,]  3.344428e-04  2.993975e-05  4.224564e-04  8.719104e-04  1.522071e-03
#>  [4,] -2.962267e-05 -7.558189e-06 -7.504791e-05 -1.580777e-04 -2.858419e-04
#>  [5,] -7.883790e-05 -1.063490e-05 -1.278824e-04 -2.245942e-04 -3.790118e-04
#>  [6,] -2.223794e-05 -7.084137e-06 -6.388759e-05 -1.623798e-04 -3.059382e-04
#>  [7,]  1.181669e-04  8.388235e-06  1.362948e-04  2.544331e-04  4.223138e-04
#>  [8,]  7.628117e-05  6.341969e-06  1.011906e-04  2.055785e-04  3.554451e-04
#>  [9,] -1.335484e-04 -1.842509e-05 -2.026006e-04 -4.091445e-04 -7.081421e-04
#> [10,]  1.228629e-04  8.093429e-06  1.284900e-04  2.111555e-04  3.423293e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.01039449 -0.05057773 -0.00489247 -0.03493585 -0.04918988 -0.02271906
#> [2,]  0.01039449  0.05057773  0.00489247  0.03493585  0.04918988  0.02271906
#>              [,7]         [,8]        [,9]       [,10]
#> [1,] -0.003426173  0.002470992 -0.03407329 -0.01706028
#> [2,]  0.003426173 -0.002470992  0.03407329  0.01706028
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.32145552 0.76705656 0.12160861 0.46665707 0.80930912 0.43020641
#>  [7] 0.17153129 0.02581755 0.58217442 0.50736461
#> 
#> $B[[2]]
#> [1]  0.1015390 -0.2090853
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0005210130  0.0012521243  0.0020539637 -0.0002103292 -0.0005401426
#>  [6] -0.0001405609  0.0007199627  0.0003444188 -0.0009980240  0.0009487914
#> 
#> $vB[[2]]
#> [1] -0.06335896  0.06335896
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [2,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#>  [3,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#>  [4,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#>  [5,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#>  [6,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#>  [7,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#>  [8,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [9,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [10,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [11,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [12,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [13,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [14,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [15,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [16,] 0.0411 0.2300 0.1370 0.1335 0.2137 0.1526 0.0775 0.1196 0.0903 0.0689
#> [17,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [18,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [19,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [20,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [21,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [22,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [23,] 0.0412 0.2058 0.1546 0.2671 0.3141 0.2904 0.3531 0.5079 0.4639 0.1859
#> [24,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [25,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [26,] 0.0790 0.6609 0.5002 0.2583 0.1650 0.4347 0.4515 0.4579 0.3366 0.4000
#> [27,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [28,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [29,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [30,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [31,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [32,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [33,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [34,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [35,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [36,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [37,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [38,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#> [39,] 0.0163 0.2822 0.3691 0.3750 0.3927 0.3308 0.1085 0.1139 0.3446 0.5441
#> [40,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [2,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#>  [3,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#>  [4,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#>  [5,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#>  [6,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#>  [7,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#>  [8,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [9,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [10,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [11,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [12,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [13,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [14,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [15,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [16,] 0.2071 0.0277 0.2975 0.2836 0.3353 0.3622 0.3202 0.3452 0.3562 0.3892
#> [17,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [18,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [19,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [20,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [21,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [22,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [23,] 0.4474 0.1135 0.4079 0.5400 0.4786 0.4332 0.6113 0.5091 0.4606 0.7243
#> [24,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [25,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [26,] 0.5325 0.0707 0.9010 0.9939 0.3689 0.1012 0.0248 0.2318 0.3981 0.2259
#> [27,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [28,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [29,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [30,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [31,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [32,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [33,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [34,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [35,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [36,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [37,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [38,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#> [39,] 0.6470 0.0198 0.7276 0.7894 0.8264 0.8697 0.7836 0.7140 0.5698 0.2908
#> [40,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [2,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#>  [3,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#>  [4,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#>  [5,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#>  [6,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#>  [7,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#>  [8,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [9,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [10,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [11,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [12,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [13,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [14,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [15,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [16,] 0.6622 0.9254 0.0604 1.0000 0.8528 0.6297 0.5250 0.4012 0.2901 0.2007
#> [17,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [18,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [19,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [20,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [21,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [22,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [23,] 0.8987 0.8826 0.0518 0.9201 0.8005 0.6033 0.2120 0.2866 0.4033 0.2803
#> [24,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [25,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [26,] 0.5247 0.6898 0.0352 0.8316 0.4326 0.3741 0.5756 0.8043 0.7963 0.7174
#> [27,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [28,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [29,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [30,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [31,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [32,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [33,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [34,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [35,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [36,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [37,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [38,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#> [39,] 0.4636 0.6409 0.0202 0.7405 0.8069 0.8420 1.0000 0.9536 0.6755 0.3905
#> [40,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [2,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#>  [3,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#>  [4,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#>  [5,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#>  [6,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#>  [7,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#>  [8,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [9,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [10,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [11,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [12,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [13,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [14,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [15,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [16,] 0.3356 0.4799 0.6147 0.0525 0.6246 0.4973 0.3492 0.2662 0.3137 0.4282
#> [17,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [18,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [19,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [20,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [21,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [22,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [23,] 0.3087 0.3550 0.2545 0.0232 0.1432 0.5869 0.6431 0.5826 0.4286 0.4894
#> [24,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [25,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [26,] 0.7056 0.8148 0.7601 0.1660 0.6034 0.4554 0.4729 0.4478 0.3722 0.4693
#> [27,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [28,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [29,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [30,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [31,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [32,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [33,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [34,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [35,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [36,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [37,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [38,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#> [39,] 0.1249 0.3629 0.6356 0.0386 0.8116 0.7664 0.5417 0.2614 0.1723 0.2814
#> [40,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [2,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#>  [3,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#>  [4,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#>  [5,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#>  [6,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#>  [7,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#>  [8,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [9,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [10,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [11,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [12,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [13,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [14,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [15,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [16,] 0.4262 0.3511 0.2458 0.1259 0.0489 0.0327 0.0181 0.0217 0.0038 0.0019
#> [17,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [18,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [19,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [20,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [21,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [22,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [23,] 0.5777 0.4315 0.2640 0.1794 0.0646 0.0772 0.0798 0.0376 0.0143 0.0272
#> [24,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [25,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [26,] 0.3839 0.0768 0.1467 0.0777 0.1330 0.0469 0.0193 0.0298 0.0390 0.0294
#> [27,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [28,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [29,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [30,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [31,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [32,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [33,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [34,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [35,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [36,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [37,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [38,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#> [39,] 0.2764 0.1985 0.1502 0.1219 0.0752 0.0493 0.0027 0.0077 0.0026 0.0031
#> [40,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [2,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#>  [3,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#>  [4,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#>  [5,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#>  [6,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#>  [7,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#>  [8,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [9,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [10,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [11,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [12,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [13,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [14,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [15,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [16,] 0.0065 0.0132 0.0108 0.0050 0.0085 0.0385 0.0044 0.0611 0.1117 0.1237
#> [17,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [18,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [19,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [20,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [21,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [22,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [23,] 0.0127 0.0166 0.0095 0.0225 0.0098 0.1124 0.0085 0.1787 0.2407 0.2682
#> [24,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [25,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [26,] 0.0175 0.0249 0.0141 0.0073 0.0025 0.0226 0.0101 0.0771 0.2678 0.5664
#> [27,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [28,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [29,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [30,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [31,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [32,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [33,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [34,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [35,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [36,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [37,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [38,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#> [39,] 0.0083 0.0020 0.0084 0.0108 0.0083 0.1444 0.0033 0.1487 0.1484 0.2442
#> [40,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5111539 0.9120274 0.4571811 0.7947515 0.8968219 0.7445604 0.3556621
#>  [2,] 0.4150017 0.8810740 0.3497133 0.7858574 0.8996269 0.6311342 0.2687270
#>  [3,] 0.4569045 0.8748129 0.4184527 0.7787097 0.9108781 0.6525199 0.3705007
#>  [4,] 0.4120368 0.8778312 0.3316782 0.7283557 0.8734885 0.6021785 0.2201557
#>  [5,] 0.3608968 0.8828582 0.2976763 0.7084741 0.8693636 0.5998171 0.2350033
#>  [6,] 0.4677033 0.8756143 0.4065047 0.7582836 0.8614402 0.6891040 0.3432716
#>  [7,] 0.3016413 0.8213727 0.2124724 0.5645186 0.7554675 0.4677440 0.1599997
#>  [8,] 0.3338890 0.8027795 0.2131971 0.6240160 0.7862063 0.4124629 0.1636556
#>  [9,] 0.4877421 0.8455119 0.4503634 0.7548938 0.8908960 0.6573224 0.4153146
#> [10,] 0.4889945 0.8491908 0.4335262 0.7605155 0.8890810 0.6430980 0.3718580
#> [11,] 0.4463036 0.8785125 0.3760687 0.7230497 0.8673397 0.6488352 0.3027713
#> [12,] 0.4530574 0.8035178 0.3909871 0.6943865 0.8099544 0.5849239 0.3783870
#> [13,] 0.5194575 0.9278660 0.4795638 0.8377798 0.9251293 0.7886434 0.3561880
#> [14,] 0.4892291 0.9047403 0.4125572 0.7967938 0.8987040 0.7257475 0.3114671
#> [15,] 0.4975511 0.8782028 0.4417029 0.7709004 0.8652702 0.6928438 0.3865400
#> [16,] 0.4069995 0.8471101 0.3129828 0.6409435 0.7916582 0.5701164 0.2494348
#> [17,] 0.4052316 0.8962445 0.3094981 0.7139941 0.8608795 0.6545758 0.2148687
#> [18,] 0.3823508 0.8214686 0.2861483 0.6372055 0.8019111 0.5017923 0.2094777
#> [19,] 0.5173857 0.9045229 0.5161360 0.8405793 0.9353488 0.7481753 0.4663599
#> [20,] 0.4575219 0.9295398 0.3708139 0.8099172 0.9250702 0.7196980 0.2347055
#> [21,] 0.4665977 0.8743424 0.4256620 0.7820947 0.9027941 0.6588541 0.3870421
#> [22,] 0.3885592 0.8557634 0.3298952 0.7250988 0.8558735 0.5897673 0.2852121
#> [23,] 0.3541823 0.8492648 0.2565628 0.6415930 0.8111862 0.5351928 0.2132155
#> [24,] 0.4051862 0.7674405 0.3438697 0.6619194 0.8372120 0.4946448 0.3384686
#> [25,] 0.3801989 0.8629705 0.3015143 0.7087470 0.8660601 0.5389676 0.2209964
#> [26,] 0.3298938 0.8307952 0.2132208 0.6176149 0.8033345 0.4307152 0.1594811
#> [27,] 0.5019439 0.8967573 0.4587684 0.8023055 0.8921337 0.7294245 0.3871959
#> [28,] 0.4328034 0.8855542 0.3137913 0.6805093 0.8183130 0.6158871 0.2196760
#> [29,] 0.4965665 0.8372952 0.4060816 0.6932662 0.8000741 0.6274082 0.3553676
#> [30,] 0.4329081 0.8689336 0.3889766 0.7602514 0.8961521 0.6158601 0.3082832
#> [31,] 0.5139146 0.8935244 0.4804910 0.7912628 0.8915619 0.7447153 0.4149215
#> [32,] 0.4958689 0.8432860 0.4547007 0.7512763 0.8576917 0.6635965 0.4187320
#> [33,] 0.3492056 0.8544824 0.2660007 0.6424105 0.8134056 0.5404128 0.2132071
#> [34,] 0.3721621 0.8930962 0.2671804 0.6732142 0.8469016 0.6101084 0.1789859
#> [35,] 0.4623328 0.8478034 0.4119996 0.7731409 0.8730091 0.6239976 0.4027479
#> [36,] 0.4714259 0.8703769 0.4418617 0.7713049 0.8811499 0.6962433 0.3866609
#> [37,] 0.3647105 0.8812480 0.2546629 0.6520681 0.8318079 0.5867808 0.1733255
#> [38,] 0.4302845 0.9074025 0.3959283 0.8087348 0.9212101 0.6758858 0.2869012
#> [39,] 0.3601404 0.8950889 0.2952037 0.7557024 0.8925814 0.5796153 0.2010713
#> [40,] 0.5298887 0.8444587 0.4910855 0.7606169 0.8514528 0.7004110 0.4495105
#>             [,8]      [,9]     [,10]
#>  [1,] 0.28522930 0.8206371 0.6547790
#>  [2,] 0.17243524 0.7435201 0.5289031
#>  [3,] 0.18733241 0.6791680 0.5629847
#>  [4,] 0.18440538 0.7296996 0.4846993
#>  [5,] 0.16581193 0.7179100 0.4861110
#>  [6,] 0.26321776 0.7768796 0.6027754
#>  [7,] 0.13665830 0.6050643 0.3633442
#>  [8,] 0.12167367 0.6214753 0.3424250
#>  [9,] 0.24827819 0.6911526 0.5846386
#> [10,] 0.23102340 0.6914450 0.5696861
#> [11,] 0.23272862 0.7481576 0.5453808
#> [12,] 0.24338734 0.6668346 0.5414423
#> [13,] 0.29496460 0.8578035 0.6704800
#> [14,] 0.23270193 0.8154830 0.6018191
#> [15,] 0.30793997 0.7856176 0.6268375
#> [16,] 0.20566198 0.7077924 0.5077563
#> [17,] 0.19150988 0.7740568 0.5117874
#> [18,] 0.17129355 0.6629412 0.4196814
#> [19,] 0.26820110 0.7888629 0.6479367
#> [20,] 0.17199807 0.8149639 0.5479840
#> [21,] 0.20749665 0.7232593 0.5830740
#> [22,] 0.15555341 0.7152813 0.5065510
#> [23,] 0.16171472 0.6707926 0.4372996
#> [24,] 0.16785386 0.5375468 0.4512202
#> [25,] 0.14872709 0.6826434 0.4614731
#> [26,] 0.09798526 0.6023161 0.3555578
#> [27,] 0.29504788 0.8106335 0.6468723
#> [28,] 0.21711523 0.7612946 0.5109809
#> [29,] 0.27956953 0.7407149 0.5810589
#> [30,] 0.19432448 0.7043223 0.5239057
#> [31,] 0.30680199 0.8038919 0.6488962
#> [32,] 0.32332005 0.7562763 0.6089877
#> [33,] 0.15071930 0.6597290 0.4516434
#> [34,] 0.15959643 0.7450020 0.4672712
#> [35,] 0.26077664 0.7282710 0.5500499
#> [36,] 0.28024897 0.7819344 0.6193974
#> [37,] 0.15478515 0.7299795 0.4497855
#> [38,] 0.19802888 0.7607641 0.5793895
#> [39,] 0.13778422 0.7166470 0.4960821
#> [40,] 0.37569633 0.7773540 0.6373311
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5593455 0.4406545
#>  [2,] 0.5644331 0.4355669
#>  [3,] 0.5589612 0.4410388
#>  [4,] 0.5637776 0.4362224
#>  [5,] 0.5645423 0.4354577
#>  [6,] 0.5597893 0.4402107
#>  [7,] 0.5622216 0.4377784
#>  [8,] 0.5640328 0.4359672
#>  [9,] 0.5561225 0.4438775
#> [10,] 0.5574849 0.4425151
#> [11,] 0.5607072 0.4392928
#> [12,] 0.5559203 0.4440797
#> [13,] 0.5605510 0.4394490
#> [14,] 0.5626222 0.4373778
#> [15,] 0.5574724 0.4425276
#> [16,] 0.5603700 0.4396300
#> [17,] 0.5654173 0.4345827
#> [18,] 0.5616139 0.4383861
#> [19,] 0.5567779 0.4432221
#> [20,] 0.5667085 0.4332915
#> [21,] 0.5590725 0.4409275
#> [22,] 0.5628881 0.4371119
#> [23,] 0.5632435 0.4367565
#> [24,] 0.5565495 0.4434505
#> [25,] 0.5636317 0.4363683
#> [26,] 0.5639743 0.4360257
#> [27,] 0.5586218 0.4413782
#> [28,] 0.5627934 0.4372066
#> [29,] 0.5562904 0.4437096
#> [30,] 0.5606784 0.4393216
#> [31,] 0.5571066 0.4428934
#> [32,] 0.5554948 0.4445052
#> [33,] 0.5627329 0.4372671
#> [34,] 0.5662595 0.4337405
#> [35,] 0.5581213 0.4418787
#> [36,] 0.5583884 0.4416116
#> [37,] 0.5659388 0.4340612
#> [38,] 0.5629721 0.4370279
#> [39,] 0.5661273 0.4338727
#> [40,] 0.5535685 0.4464315
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]     [,2]        [,3]      [,4]     [,5]         [,6]
#>  [1,]  0.044623026 2.338644 -0.17169597 1.3538079 2.162400  1.069808171
#>  [2,] -0.343326137 2.002640 -0.62029951 1.3001332 2.193086  0.537085595
#>  [3,] -0.172810892 1.944201 -0.32912854 1.2581627 2.324403  0.630134053
#>  [4,] -0.355551899 1.972051 -0.70060451 0.9862962 1.932162  0.414550342
#>  [5,] -0.571473771 2.019780 -0.85838793 0.8879847 1.895343  0.404703192
#>  [6,] -0.129366755 1.951539 -0.37843361 1.1432923 1.827303  0.795933599
#>  [7,] -0.839494345 1.525676 -1.31008655 0.2595212 1.127989 -0.129203331
#>  [8,] -0.690647782 1.403758 -1.30576095 0.5066293 1.302208 -0.353793240
#>  [9,] -0.049041533 1.699825 -0.19920248 1.1248855 2.099927  0.651384476
#> [10,] -0.044029249 1.728268 -0.26747880 1.1555079 2.081388  0.588835845
#> [11,] -0.215617252 1.978418 -0.50626836 0.9596400 1.877639  0.613922980
#> [12,] -0.188324930 1.408427 -0.44316471 0.8207074 1.449714  0.343020035
#> [13,]  0.077869275 2.554362 -0.08179023 1.6418005 2.514171  1.316767538
#> [14,] -0.043090089 2.251041 -0.35340382 1.3663748 2.182906  0.973153098
#> [15,] -0.009795704 1.975520 -0.23425374 1.2134024 1.859771  0.813448063
#> [16,] -0.376383458 1.712113 -0.78621126 0.5794617 1.334949  0.282326048
#> [17,] -0.383713349 2.156176 -0.80246678 0.9148624 1.822614  0.639214063
#> [18,] -0.479582245 1.526329 -0.91416485 0.5632556 1.398282  0.007169071
#> [19,]  0.069570974 2.248522  0.06456625 1.6625449 2.671914  1.088904183
#> [20,] -0.170322837 2.579641 -0.52872689 1.4494719 2.513317  0.942964256
#> [21,] -0.133808319 1.939912 -0.29957258 1.2779152 2.228663  0.658191924
#> [22,] -0.453372517 1.780539 -0.70865923 0.9698961 1.781431  0.363003335
#> [23,] -0.600705652 1.728847 -1.06391108 0.5822851 1.457736  0.141004222
#> [24,] -0.383901895 1.193915 -0.64609666 0.6718597 1.637629 -0.021421586
#> [25,] -0.488704249 1.840184 -0.84009749 0.8893063 1.866563  0.156187239
#> [26,] -0.708665231 1.591273 -1.30561912 0.4794369 1.407267 -0.278933599
#> [27,]  0.007775791 2.161702 -0.16530169 1.4007664 2.112723  0.991704486
#> [28,] -0.270422367 2.046113 -0.78245358 0.7561132 1.504959  0.472127094
#> [29,] -0.013734404 1.638239 -0.38018793 0.8154336 1.386758  0.521113872
#> [30,] -0.269995848 1.891562 -0.45161593 1.1540585 2.155183  0.472013214
#> [31,]  0.055672668 2.127258 -0.07807548 1.3325542 2.106796  1.070622765
#> [32,] -0.016524623 1.682884 -0.18169554 1.1054306 1.796249  0.679363084
#> [33,] -0.622532884 1.770199 -1.01500893 0.5858420 1.472293  0.162004413
#> [34,] -0.522952448 2.122765 -1.00897573 0.7227588 1.710504  0.447767794
#> [35,] -0.150954618 1.717476 -0.35570544 1.2261321 1.927830  0.506551094
#> [36,] -0.114421248 1.904296 -0.23360988 1.2156943 2.003364  0.829472270
#> [37,] -0.554976712 2.004302 -1.07389598 0.6281422 1.598495  0.350673105
#> [38,] -0.280690414 2.282324 -0.42245991 1.4418100 2.458903  0.734928269
#> [39,] -0.574754984 2.143810 -0.87024308 1.1292604 2.117384  0.321194296
#> [40,]  0.119697668 1.691785 -0.03566166 1.1560647 1.746042  0.849255573
#>             [,7]       [,8]      [,9]       [,10]
#>  [1,] -0.5942421 -0.9186684 1.5206699  0.64011305
#>  [2,] -1.0010911 -1.5684656 1.0643456  0.11574135
#>  [3,] -0.5300694 -1.4674375 0.7499510  0.25328423
#>  [4,] -1.2647591 -1.4867809 0.9930993 -0.06122188
#>  [5,] -1.1802717 -1.6156047 0.9341182 -0.05557040
#>  [6,] -0.6487486 -1.0293107 1.2475740  0.41704288
#>  [7,] -1.6582304 -1.8433269 0.4266117 -0.56087870
#>  [8,] -1.6312764 -1.9766756 0.4958150 -0.65250613
#>  [9,] -0.3420376 -1.1078165 0.8055133  0.34184471
#> [10,] -0.5242543 -1.2025415 0.8068837  0.28057037
#> [11,] -0.8341355 -1.1929675 1.0888101  0.18202421
#> [12,] -0.4964002 -1.1341973 0.6939028  0.16615030
#> [13,] -0.5919481 -0.8713926 1.7971651  0.71035681
#> [14,] -0.7932695 -1.1931170 1.4860389  0.41305034
#> [15,] -0.4618797 -0.8097678 1.2987091  0.51867347
#> [16,] -1.1016290 -1.3512751 0.8846864  0.03102771
#> [17,] -1.2958238 -1.4402291 1.2313616  0.04715852
#> [18,] -1.3280767 -1.5764872 0.6764290 -0.32408129
#> [19,] -0.1347642 -1.0037687 1.3180846  0.60998189
#> [20,] -1.1819294 -1.5715322 1.4825926  0.19252838
#> [21,] -0.4597627 -1.3400815 0.9606868  0.33540538
#> [22,] -0.9187527 -1.6916924 0.9211742  0.02620531
#> [23,] -1.3056511 -1.6455247 0.7117721 -0.25212898
#> [24,] -0.6701260 -1.6009143 0.1504703 -0.19574191
#> [25,] -1.2598693 -1.7446198 0.7659464 -0.15441364
#> [26,] -1.6620939 -2.2198138 0.4151250 -0.59469751
#> [27,] -0.4591145 -0.8709923 1.4541315  0.60531913
#> [28,] -1.2675556 -1.2825573 1.1597901  0.04393047
#> [29,] -0.5955276 -0.9465979 1.0496876  0.32712185
#> [30,] -0.8081577 -1.4221518 0.8679659  0.09569586
#> [31,] -0.3436567 -0.8151131 1.4107989  0.61419091
#> [32,] -0.3279808 -0.7385557 1.1323713  0.44305903
#> [33,] -1.3057009 -1.7289706 0.6620869 -0.19403282
#> [34,] -1.5232334 -1.6612339 1.0721311 -0.13110277
#> [35,] -0.3940283 -1.0419358 0.9858681  0.20087249
#> [36,] -0.4613700 -0.9432270 1.2769751  0.48699110
#> [37,] -1.5622400 -1.6975528 0.9945185 -0.20153736
#> [38,] -0.9104822 -1.3986597 1.1568731  0.32026772
#> [39,] -1.3796122 -1.8338167 0.9278898 -0.01567197
#> [40,] -0.2026487 -0.5078557 1.2503128  0.56379914
#> 
#> $pre[[3]]
#>                [,1]       [,2]
#>  [1,] -0.0056739601 -0.2441802
#>  [2,]  0.0221551531 -0.2370184
#>  [3,]  0.0046539279 -0.2322932
#>  [4,]  0.0207144512 -0.2357934
#>  [5,]  0.0231491222 -0.2364685
#>  [6,] -0.0001465322 -0.2404533
#>  [7,]  0.0252194949 -0.2249638
#>  [8,]  0.0309658041 -0.2265796
#>  [9,] -0.0087684980 -0.2342083
#> [10,] -0.0026101054 -0.2335710
#> [11,]  0.0048514234 -0.2391813
#> [12,] -0.0060627245 -0.2306837
#> [13,] -0.0027292907 -0.2461277
#> [14,]  0.0075767317 -0.2442343
#> [15,] -0.0098193329 -0.2407293
#> [16,]  0.0070521406 -0.2356115
#> [17,]  0.0211670032 -0.2420108
#> [18,]  0.0173900279 -0.2303246
#> [19,] -0.0116432995 -0.2397387
#> [20,]  0.0246600501 -0.2437744
#> [21,]  0.0009909384 -0.2364079
#> [22,]  0.0177125596 -0.2351791
#> [23,]  0.0213247911 -0.2330114
#> [24,]  0.0055764492 -0.2215936
#> [25,]  0.0233969663 -0.2325173
#> [26,]  0.0310346874 -0.2262730
#> [27,] -0.0070082339 -0.2425786
#> [28,]  0.0122623772 -0.2402442
#> [29,] -0.0114294352 -0.2375497
#> [30,]  0.0108877322 -0.2330282
#> [31,] -0.0127743789 -0.2422019
#> [32,] -0.0153406465 -0.2382383
#> [33,]  0.0208684963 -0.2313922
#> [34,]  0.0261456790 -0.2404602
#> [35,] -0.0019074936 -0.2354482
#> [36,] -0.0059495073 -0.2405736
#> [37,]  0.0259660021 -0.2393345
#> [38,]  0.0160915835 -0.2371413
#> [39,]  0.0315036875 -0.2345643
#> [40,] -0.0240434992 -0.2391431
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.5593455  0.5593455
#>  [2,]  0.4355669 -0.4355669
#>  [3,] -0.5589612  0.5589612
#>  [4,] -0.5637776  0.5637776
#>  [5,]  0.4354577 -0.4354577
#>  [6,]  0.4402107 -0.4402107
#>  [7,]  0.4377784 -0.4377784
#>  [8,]  0.4359672 -0.4359672
#>  [9,]  0.4438775 -0.4438775
#> [10,] -0.5574849  0.5574849
#> [11,] -0.5607072  0.5607072
#> [12,]  0.4440797 -0.4440797
#> [13,] -0.5605510  0.5605510
#> [14,]  0.4373778 -0.4373778
#> [15,] -0.5574724  0.5574724
#> [16,]  0.4396300 -0.4396300
#> [17,]  0.4345827 -0.4345827
#> [18,]  0.4383861 -0.4383861
#> [19,] -0.5567779  0.5567779
#> [20,] -0.5667085  0.5667085
#> [21,]  0.4409275 -0.4409275
#> [22,]  0.4371119 -0.4371119
#> [23,]  0.4367565 -0.4367565
#> [24,]  0.4434505 -0.4434505
#> [25,] -0.5636317  0.5636317
#> [26,]  0.4360257 -0.4360257
#> [27,]  0.4413782 -0.4413782
#> [28,] -0.5627934  0.5627934
#> [29,] -0.5562904  0.5562904
#> [30,]  0.4393216 -0.4393216
#> [31,] -0.5571066  0.5571066
#> [32,] -0.5554948  0.5554948
#> [33,]  0.4372671 -0.4372671
#> [34,]  0.4337405 -0.4337405
#> [35,] -0.5581213  0.5581213
#> [36,]  0.4416116 -0.4416116
#> [37,]  0.4340612 -0.4340612
#> [38,]  0.4370279 -0.4370279
#> [39,]  0.4338727 -0.4338727
#> [40,] -0.5535685  0.5535685
#> 
#> $L
#> [1] 0.6962310 0.6829717 0.7029464 0.6904937 0.7111706 0.6736834
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609