Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1           V10         V11          V12          V13
#>  [1,] -0.1990430 -0.1272603468 -0.05857201  0.002273123 -0.082360116
#>  [2,] -0.2457515 -0.1290683028 -0.10094777 -0.125402697 -0.088618963
#>  [3,] -0.3430021 -0.0122599717 -0.10033113 -0.059300001 -0.080126816
#>  [4,] -0.2516408  0.0041384321 -0.00266615 -0.043444526 -0.070787982
#>  [5,] -0.3632608 -0.0881293830  0.01160178 -0.009068839 -0.025772979
#>  [6,] -0.1744318 -0.1040953462 -0.16656105 -0.191683551 -0.189406303
#>  [7,] -0.3647393 -0.0006391614 -0.08185575 -0.107315624  0.007495583
#>  [8,] -0.3374495 -0.1416891114 -0.10092076 -0.029240185 -0.002272083
#>  [9,] -0.3901354 -0.0851414392 -0.08271275 -0.126353995 -0.087054107
#> [10,] -0.3294899 -0.2138054999 -0.18067852 -0.132353783  0.026845932
#>                 V14          V15          V16         V17         V18
#>  [1,] -0.0009709522 -0.037005638 -0.042646693 -0.11433486 -0.04259733
#>  [2,] -0.0898009725  0.048559875 -0.100237796 -0.05567557 -0.13869080
#>  [3,]  0.0368029443  0.054543745  0.139127619  0.08230099  0.10913418
#>  [4,] -0.0518370186 -0.092619330  0.059247087  0.04961684  0.16054568
#>  [5,] -0.0372563356 -0.039771996 -0.061531988  0.03717227 -0.01351758
#>  [6,] -0.0265492650 -0.027729142 -0.063980112 -0.08059097 -0.01238437
#>  [7,] -0.0355969079 -0.055657626  0.094079878 -0.03337741  0.07953701
#>  [8,] -0.0093359806 -0.058971634 -0.081562635 -0.12457680 -0.11793771
#>  [9,] -0.0274498653 -0.148857007 -0.004061222 -0.02022905  0.11058774
#> [10,] -0.1266098127 -0.007049609 -0.020349722  0.08614948 -0.08511960
#>               V19         V2         V20           V21         V22          V23
#>  [1,] -0.02981473 -0.3026985 -0.02311257  0.0001117072  0.03915461 -0.084108907
#>  [2,] -0.08673937 -0.2686271  0.12158163  0.1411121986  0.10943437  0.081300374
#>  [3,]  0.01001128 -0.2063763  0.10672023  0.0864030871  0.17369898  0.095242402
#>  [4,]  0.11594403 -0.2223840  0.07215049  0.1818794397  0.11459234  0.224585062
#>  [5,]  0.01485162 -0.3981909  0.11683322  0.1176511772  0.19091699  0.228026554
#>  [6,]  0.03223721 -0.2566900 -0.11165198 -0.0638591314 -0.13565087  0.031894382
#>  [7,]  0.09180981 -0.4036659  0.24677871  0.2110068256  0.23220549  0.115134405
#>  [8,]  0.04305997 -0.2822608  0.04422726 -0.0221065198 -0.03749966 -0.004738738
#>  [9,]  0.13392010 -0.3616391  0.03086542  0.0525480195  0.11155206  0.134771367
#> [10,] -0.07409010 -0.2720509  0.09553235  0.0695178723  0.02112818  0.099567688
#>                V24         V25         V26         V27         V28         V29
#>  [1,] -0.004811192  0.05598915  0.05567638  0.18280903  0.08009059  0.08417882
#>  [2,]  0.171970518  0.20817717  0.14720919  0.07876545  0.08646151  0.22239018
#>  [3,]  0.317121921  0.15998142  0.31753430  0.24155696  0.08375753  0.22417048
#>  [4,]  0.218895334  0.18061719  0.16244546  0.16805215  0.21602657  0.10479508
#>  [5,]  0.148248175  0.22855085  0.27589928  0.16733074  0.30006332  0.16868032
#>  [6,] -0.033913824 -0.04525368  0.04608331 -0.07791272 -0.10894483 -0.06013021
#>  [7,]  0.126554255  0.21532485  0.26636426  0.17575885  0.13578434  0.16031431
#>  [8,]  0.103101639  0.09595519  0.11770698  0.09103169  0.16387907  0.09013212
#>  [9,]  0.101589366  0.15813460  0.08548657  0.17298591  0.06951875  0.24639506
#> [10,] -0.029160191  0.02624749 -0.02362729  0.11038353  0.16555825  0.09294189
#>               V3        V30           V31          V32         V33         V34
#>  [1,] -0.2286589 0.09793086 -0.0493857035  0.009818332 -0.03400705  0.06536625
#>  [2,] -0.2270917 0.12069068  0.0401567874 -0.032109534  0.03755462 -0.01843592
#>  [3,] -0.2256274 0.13122898  0.0081638881  0.035272634 -0.01385054  0.02122202
#>  [4,] -0.2383625 0.17660851  0.1364790820  0.082471520  0.02131276  0.02257981
#>  [5,] -0.2710644 0.21644364  0.1298429422  0.038274971  0.06273064 -0.07914594
#>  [6,] -0.1545935 0.03622257 -0.0689512025  0.016430827 -0.07155941 -0.03909952
#>  [7,] -0.2623900 0.23236994  0.1966763116  0.053614234  0.09033195  0.04902339
#>  [8,] -0.3605834 0.03244915 -0.0007537759 -0.009281937  0.03951764 -0.06443779
#>  [9,] -0.2477632 0.11790868  0.1760293145  0.021531336  0.07210549  0.13826655
#> [10,] -0.3165682 0.08316431 -0.0625827860 -0.057099836 -0.09962642 -0.01896870
#>                V35          V36          V37         V38          V39
#>  [1,]  0.030881454 -0.023861845  0.058748040  0.06009210 -0.085543613
#>  [2,]  0.075309228 -0.037294258 -0.032085198  0.01020720 -0.019187203
#>  [3,]  0.000334696  0.059151526 -0.051519316 -0.07848206 -0.032576305
#>  [4,]  0.058918023 -0.002293271  0.042221323  0.04537388  0.073192151
#>  [5,]  0.019593990  0.033073763  0.009045642 -0.07315415 -0.142608321
#>  [6,] -0.022700308  0.011885056 -0.116075889 -0.03828239  0.033538338
#>  [7,]  0.105831078 -0.025896491 -0.012859097  0.02013095 -0.004649235
#>  [8,] -0.103811153  0.002767114 -0.023377094 -0.07266074 -0.104882361
#>  [9,] -0.025647955  0.020216363 -0.051697387 -0.10986092 -0.033378527
#> [10,]  0.037001117  0.005582895  0.033762125  0.03888078  0.004559016
#>               V4          V40          V41           V42         V43
#>  [1,] -0.2716958 -0.016161367  0.053574314 -0.0286536891 -0.09664518
#>  [2,] -0.2272347 -0.013687316  0.011717995  0.0156445560 -0.09018426
#>  [3,] -0.2136492  0.016176647 -0.133040631 -0.0171327773 -0.06618684
#>  [4,] -0.2942506  0.028945991 -0.078497922  0.0008012918 -0.13568986
#>  [5,] -0.2566201 -0.082227499 -0.010586856 -0.0407720369 -0.02650657
#>  [6,] -0.1333836  0.005457838 -0.169879103 -0.2344591102 -0.11578528
#>  [7,] -0.3232028  0.002735716 -0.074295160 -0.0399163993 -0.06922888
#>  [8,] -0.2629415 -0.143565879 -0.006286813 -0.1232602003 -0.14392301
#>  [9,] -0.2891260 -0.067705944 -0.051156602 -0.0523034232 -0.17226060
#> [10,] -0.1564344 -0.136806439 -0.099832725 -0.0526172115 -0.13276607
#>               V44         V45         V46         V47         V48        V49
#>  [1,] -0.01221851 -0.12620803 -0.03640554 -0.19231491 -0.21548707 -0.2492954
#>  [2,] -0.18144769 -0.18113426 -0.12185149 -0.13442041 -0.19647384 -0.2905589
#>  [3,] -0.06505280 -0.16040828 -0.10726704 -0.09477201 -0.09043646 -0.2935321
#>  [4,] -0.01525835 -0.08201715 -0.19563585 -0.19956931 -0.28766703 -0.2156551
#>  [5,] -0.16718555 -0.05191611 -0.17775294 -0.16197763 -0.30332215 -0.2707226
#>  [6,] -0.08882260 -0.16127582 -0.08501123 -0.26891881 -0.15719689 -0.1514252
#>  [7,] -0.04483446 -0.02035186 -0.14929174 -0.14270810 -0.29518407 -0.2994210
#>  [8,] -0.20136506 -0.19429887 -0.07939526 -0.18020681 -0.13224779 -0.1875200
#>  [9,] -0.19622187 -0.07436977 -0.06147934 -0.29115222 -0.22910232 -0.2993552
#> [10,] -0.18801745 -0.10031840 -0.10528466 -0.10844651 -0.25901240 -0.2541954
#>               V5        V50        V51        V52        V53        V54
#>  [1,] -0.1086065 -0.3386787 -0.2161363 -0.3608610 -0.3184970 -0.2522083
#>  [2,] -0.2525187 -0.1732113 -0.1982103 -0.2663371 -0.2507509 -0.2305886
#>  [3,] -0.2674706 -0.2146047 -0.3189693 -0.3475292 -0.2504050 -0.2066658
#>  [4,] -0.1806753 -0.3980979 -0.2587380 -0.3309310 -0.2576257 -0.3878211
#>  [5,] -0.3742282 -0.4009864 -0.4206590 -0.4274927 -0.2721935 -0.3078615
#>  [6,] -0.2228707 -0.3040254 -0.2392624 -0.3284035 -0.2670837 -0.1616182
#>  [7,] -0.2314106 -0.3521153 -0.4486491 -0.3625736 -0.4511184 -0.4476730
#>  [8,] -0.1917204 -0.2212594 -0.2452136 -0.2565574 -0.3852985 -0.2712974
#>  [9,] -0.2858211 -0.4088330 -0.4320824 -0.3125081 -0.3461377 -0.4468911
#> [10,] -0.2332888 -0.2976800 -0.3484919 -0.3422567 -0.3572648 -0.3625571
#>              V55        V56        V57        V58        V59         V6
#>  [1,] -0.3593153 -0.2831035 -0.3423400 -0.2751724 -0.2417375 -0.2282692
#>  [2,] -0.2549061 -0.3651711 -0.3262935 -0.3124639 -0.2845273 -0.1271383
#>  [3,] -0.3763003 -0.2534235 -0.3659394 -0.2838432 -0.2587093 -0.2290795
#>  [4,] -0.2830252 -0.4504402 -0.3401793 -0.3639759 -0.4240342 -0.2180116
#>  [5,] -0.3978910 -0.3463199 -0.3007957 -0.2962481 -0.3146378 -0.2012424
#>  [6,] -0.2372480 -0.2919692 -0.3128476 -0.1721954 -0.2299651 -0.1331138
#>  [7,] -0.4769345 -0.4367355 -0.3819090 -0.3838295 -0.3491931 -0.2655928
#>  [8,] -0.2557339 -0.3666897 -0.2370055 -0.3372738 -0.3496194 -0.2720996
#>  [9,] -0.2846798 -0.4375012 -0.3122720 -0.4042178 -0.4468330 -0.2239260
#> [10,] -0.2890901 -0.2167308 -0.3947601 -0.3574194 -0.2498216 -0.1154426
#>              V60          V7          V8          V9
#>  [1,] -0.2309158 -0.22785757 -0.09475425 -0.06607840
#>  [2,] -0.2055273 -0.20991892 -0.14543866 -0.08154400
#>  [3,] -0.2983502 -0.05698174 -0.18913437 -0.01290634
#>  [4,] -0.3735213 -0.24062113 -0.24904104 -0.03597048
#>  [5,] -0.3343805 -0.27105610 -0.12079634 -0.22551158
#>  [6,] -0.2644290 -0.19943122 -0.14713195 -0.12995998
#>  [7,] -0.4845525 -0.17847424 -0.21327583 -0.08967008
#>  [8,] -0.3025030 -0.13860164 -0.22616308 -0.12665298
#>  [9,] -0.2711000 -0.09633661 -0.14260064 -0.14837643
#> [10,] -0.3088270 -0.19802295 -0.04528358 -0.14759255
#> 
#> $W[[2]]
#>              [,1]        [,2]        [,3]        [,4]       [,5]      [,6]
#> [1,] -0.004067333  0.07954315  0.02122284  0.04478888 0.09484901 0.0446202
#> [2,] -0.080661508 -0.10393578 -0.10913070 -0.03645594 0.05285660 0.1414873
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.08887941 -0.05445818 -0.01688027 -0.01744491
#> [2,] -0.07370336  0.05249596  0.04560912 -0.03086324
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,] -3.692080e-05 -0.0002403164 -0.0002236611 -2.077272e-04 -1.261388e-04
#>  [2,] -1.677056e-04 -0.0011914398 -0.0012045283 -1.227941e-03 -1.047073e-03
#>  [3,] -2.023142e-05 -0.0001377336 -0.0001135073 -8.460945e-05 -4.880434e-05
#>  [4,]  3.666053e-05  0.0002668418  0.0002976900  3.203342e-04  3.179929e-04
#>  [5,]  8.123346e-05  0.0005766350  0.0006118717  6.306621e-04  6.062333e-04
#>  [6,]  1.051794e-04  0.0008268915  0.0008812182  9.703424e-04  8.461402e-04
#>  [7,] -3.261545e-05 -0.0002157002 -0.0002012928 -1.845812e-04 -1.412299e-04
#>  [8,]  2.617710e-04  0.0020106186  0.0021355470  2.306772e-03  2.077403e-03
#>  [9,]  2.177470e-04  0.0015957906  0.0016831700  1.768801e-03  1.635227e-03
#> [10,]  4.388568e-05  0.0003678701  0.0004096430  4.642167e-04  4.652114e-04
#>                 V14           V15           V16           V17           V18
#>  [1,] -4.434122e-05  8.646839e-07  1.131584e-04  1.635956e-04  1.589104e-04
#>  [2,] -7.240909e-04 -5.530098e-04 -2.507970e-04 -3.527209e-05 -3.351716e-05
#>  [3,] -3.671477e-05 -3.474427e-05  2.717995e-05  7.088765e-05  7.970461e-05
#>  [4,]  2.542297e-04  2.259845e-04  2.209963e-04  1.917141e-04  1.831547e-04
#>  [5,]  4.820515e-04  4.300719e-04  3.669437e-04  2.789647e-04  2.562908e-04
#>  [6,]  5.514135e-04  3.975996e-04  2.179107e-04  5.712119e-05  3.934952e-05
#>  [7,] -1.080627e-04 -9.477875e-05 -2.299700e-05  3.293561e-05  4.031592e-05
#>  [8,]  1.458092e-03  1.113436e-03  7.573029e-04  4.721783e-04  4.874627e-04
#>  [9,]  1.233730e-03  1.041383e-03  8.223132e-04  5.964141e-04  5.889382e-04
#> [10,]  3.668262e-04  3.157026e-04  3.042377e-04  2.562250e-04  2.551176e-04
#>                 V19            V2           V20           V21           V22
#>  [1,] -2.220753e-05 -3.576121e-05 -0.0002114708 -2.088248e-04 -0.0001151112
#>  [2,] -8.269524e-04 -1.743608e-04 -0.0015901066 -1.650716e-03 -0.0013866776
#>  [3,]  1.720431e-06 -1.730158e-05 -0.0001196226 -9.707715e-05 -0.0000311694
#>  [4,]  3.403901e-04  4.310822e-05  0.0004484706  4.959609e-04  0.0004646371
#>  [5,]  6.136778e-04  9.015235e-05  0.0009113257  9.810353e-04  0.0008724227
#>  [6,]  6.046695e-04  1.100659e-04  0.0010667536  1.113387e-03  0.0009985278
#>  [7,] -9.359625e-05 -3.059641e-05 -0.0002575854 -2.453534e-04 -0.0001655282
#>  [8,]  1.808601e-03  2.836448e-04  0.0029063256  3.080887e-03  0.0028171047
#>  [9,]  1.586583e-03  2.362740e-04  0.0024571590  2.603607e-03  0.0023395779
#> [10,]  4.739001e-04  4.978909e-05  0.0006388301  6.806893e-04  0.0006765140
#>                 V23           V24           V25           V26           V27
#>  [1,] -3.982160e-05 -1.839666e-05  6.803747e-05  9.178465e-05  4.635692e-05
#>  [2,] -1.170257e-03 -1.129569e-03 -9.697430e-04 -1.036445e-03 -1.248915e-03
#>  [3,]  2.070299e-05  2.108827e-05  5.720864e-05  6.169750e-05  3.367201e-05
#>  [4,]  4.378662e-04  4.381990e-04  4.528095e-04  4.915286e-04  5.319508e-04
#>  [5,]  7.794511e-04  7.662235e-04  7.567740e-04  8.252931e-04  9.131923e-04
#>  [6,]  8.469078e-04  7.723137e-04  7.024727e-04  7.717700e-04  9.183944e-04
#>  [7,] -1.002579e-04 -9.905582e-05 -6.631783e-05 -7.146108e-05 -1.201339e-04
#>  [8,]  2.546465e-03  2.474799e-03  2.326655e-03  2.506965e-03  2.850773e-03
#>  [9,]  2.107040e-03  2.085749e-03  2.011245e-03  2.182209e-03  2.451719e-03
#> [10,]  6.606962e-04  6.597051e-04  6.871488e-04  7.608299e-04  8.186718e-04
#>                 V28           V29            V3           V30           V31
#>  [1,]  1.575908e-05  1.173306e-04 -3.753516e-05  1.691364e-04  2.370312e-04
#>  [2,] -1.364400e-03 -1.009764e-03 -1.908994e-04 -8.123502e-04 -4.051023e-04
#>  [3,]  4.042810e-05  1.159612e-04 -2.035049e-05  1.260421e-04  1.575342e-04
#>  [4,]  5.852877e-04  5.391315e-04  4.419873e-05  5.008162e-04  3.980066e-04
#>  [5,]  9.999186e-04  8.668038e-04  9.373933e-05  8.057236e-04  5.831298e-04
#>  [6,]  9.861831e-04  8.073159e-04  1.298069e-04  7.033532e-04  4.323820e-04
#>  [7,] -1.265182e-04 -3.471708e-05 -3.498965e-05 -9.862310e-06  4.520873e-05
#>  [8,]  3.011762e-03  2.512088e-03  3.217860e-04  2.165980e-03  1.436702e-03
#>  [9,]  2.618707e-03  2.214945e-03  2.580494e-04  1.964689e-03  1.362956e-03
#> [10,]  8.363268e-04  7.794702e-04  5.952725e-05  7.272294e-04  5.784457e-04
#>                 V32           V33           V34           V35           V36
#>  [1,]  1.395270e-04  1.041285e-04  1.207850e-04  1.972586e-04  2.906183e-04
#>  [2,] -6.294109e-04 -6.847676e-04 -4.241563e-04 -6.050264e-05  2.172896e-04
#>  [3,]  1.003594e-04  6.045247e-05  6.720356e-05  9.314189e-05  1.312616e-04
#>  [4,]  4.050706e-04  3.688781e-04  2.846172e-04  2.179580e-04  1.618477e-04
#>  [5,]  6.404923e-04  6.256510e-04  4.689482e-04  3.312164e-04  2.054415e-04
#>  [6,]  5.441769e-04  5.189719e-04  2.289753e-04 -8.961832e-05 -2.348421e-04
#>  [7,] -1.964411e-05 -5.423967e-05 -2.576782e-05  2.638268e-05  8.931767e-05
#>  [8,]  1.653599e-03  1.579715e-03  9.759318e-04  2.982821e-04 -5.165572e-05
#>  [9,]  1.525757e-03  1.510590e-03  1.094131e-03  6.397445e-04  3.429188e-04
#> [10,]  5.546618e-04  5.206538e-04  3.714672e-04  2.457663e-04  2.265734e-04
#>                 V37           V38           V39            V4           V40
#>  [1,]  1.754916e-04  2.426436e-05 -1.791306e-05 -6.488802e-05  0.0001168434
#>  [2,] -1.080975e-04 -5.430803e-04 -5.839529e-04 -3.171626e-04 -0.0002125635
#>  [3,]  6.159887e-05  8.473573e-06  2.961983e-05 -3.862413e-05  0.0001066156
#>  [4,]  1.676911e-04  2.320543e-04  2.466012e-04  7.214551e-05  0.0002189112
#>  [5,]  2.603952e-04  4.176860e-04  4.229345e-04  1.598631e-04  0.0003160868
#>  [6,]  3.802038e-05  3.191384e-04  3.649339e-04  2.067533e-04  0.0002317171
#>  [7,]  1.196188e-05 -5.872613e-05 -3.428222e-05 -6.377317e-05  0.0000589092
#>  [8,]  4.305658e-04  1.086408e-03  1.212482e-03  5.101620e-04  0.0008129614
#>  [9,]  6.270522e-04  1.075983e-03  1.089642e-03  4.235681e-04  0.0007342199
#> [10,]  2.760961e-04  3.267065e-04  3.122691e-04  8.954887e-05  0.0003030011
#>                 V41           V42           V43           V44           V45
#>  [1,]  9.103589e-05 -4.276554e-05 -6.890368e-05 -1.225331e-04 -1.844158e-04
#>  [2,] -3.219589e-04 -7.834723e-04 -8.330186e-04 -9.930745e-04 -1.166834e-03
#>  [3,]  9.615349e-05  2.686962e-05  1.704845e-05 -2.726842e-05 -6.964838e-05
#>  [4,]  2.434510e-04  2.967768e-04  2.967650e-04  3.066857e-04  3.293730e-04
#>  [5,]  3.710405e-04  5.254575e-04  5.340382e-04  5.825809e-04  6.557405e-04
#>  [6,]  3.242756e-04  6.278716e-04  6.404790e-04  7.428619e-04  8.229259e-04
#>  [7,]  4.037630e-05 -5.376386e-05 -6.724585e-05 -1.157178e-04 -1.636409e-04
#>  [8,]  1.010193e-03  1.680029e-03  1.717485e-03  1.936086e-03  2.120053e-03
#>  [9,]  8.574685e-04  1.342501e-03  1.377297e-03  1.534502e-03  1.709623e-03
#> [10,]  3.333224e-04  4.158499e-04  4.028056e-04  4.205009e-04  4.242520e-04
#>                 V46           V47           V48           V49            V5
#>  [1,] -1.236769e-04 -7.095329e-05 -7.027007e-05 -4.327377e-05 -1.922732e-05
#>  [2,] -7.947777e-04 -5.063622e-04 -4.311979e-04 -2.459447e-04 -2.484725e-04
#>  [3,] -2.777279e-05 -1.231924e-05 -2.981610e-05 -1.414221e-05 -6.076939e-06
#>  [4,]  2.556322e-04  1.683823e-04  1.200293e-04  6.916036e-05  9.075847e-05
#>  [5,]  4.818411e-04  3.068106e-04  2.348055e-04  1.343531e-04  1.688869e-04
#>  [6,]  5.647172e-04  3.910280e-04  3.150498e-04  1.831424e-04  2.017788e-04
#>  [7,] -9.814016e-05 -5.706578e-05 -6.317975e-05 -3.183676e-05 -3.221533e-05
#>  [8,]  1.474977e-03  9.912440e-04  7.899932e-04  4.548182e-04  5.129244e-04
#>  [9,]  1.208151e-03  7.868540e-04  6.246815e-04  3.532379e-04  4.230561e-04
#> [10,]  2.926893e-04  2.073206e-04  1.583044e-04  8.783187e-05  1.228827e-04
#>                 V50           V51           V52           V53           V54
#>  [1,] -5.003259e-07 -1.867884e-05 -1.368135e-05 -7.064113e-06 -7.930047e-06
#>  [2,] -4.600381e-05 -8.785424e-05 -5.804956e-05 -3.480479e-05 -4.360835e-05
#>  [3,]  3.346796e-06 -8.779813e-06 -4.938428e-06 -2.405159e-06 -4.234096e-06
#>  [4,]  2.213467e-05  2.073314e-05  1.529210e-05  8.926593e-06  1.040762e-05
#>  [5,]  3.775970e-05  4.353281e-05  3.037051e-05  1.794035e-05  2.254574e-05
#>  [6,]  3.424895e-05  5.407886e-05  3.024634e-05  1.859643e-05  2.503833e-05
#>  [7,] -2.152688e-06 -1.504811e-05 -9.137962e-06 -5.228902e-06 -7.725402e-06
#>  [8,]  1.012968e-04  1.426269e-04  9.158428e-05  5.673524e-05  7.114110e-05
#>  [9,]  8.864940e-05  1.170423e-04  7.802533e-05  4.791426e-05  6.074172e-05
#> [10,]  2.612382e-05  2.408006e-05  1.348576e-05  9.847311e-06  1.363879e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  2.094056e-06 -4.798915e-06 -2.043340e-06 -2.480346e-06 -5.996835e-07
#>  [2,] -8.674086e-06 -3.123929e-05 -1.925182e-05 -2.400703e-05 -1.888854e-05
#>  [3,]  1.961233e-06 -2.820517e-06 -1.803616e-06 -2.280698e-07  5.261921e-07
#>  [4,]  7.220978e-06  7.929595e-06  5.687410e-06  9.288580e-06  8.417691e-06
#>  [5,]  1.165387e-05  1.695213e-05  1.189743e-05  1.720200e-05  1.530414e-05
#>  [6,]  3.216120e-06  2.076685e-05  1.049084e-05  1.557785e-05  1.363986e-05
#>  [7,]  2.607134e-07 -5.470469e-06 -3.813039e-06 -3.238814e-06 -2.264965e-06
#>  [8,]  2.124522e-05  5.361437e-05  3.200439e-05  4.371553e-05  3.849765e-05
#>  [9,]  2.443310e-05  4.501937e-05  3.009464e-05  3.993528e-05  3.546521e-05
#> [10,]  7.274579e-06  1.103257e-05  7.440314e-06  9.653117e-06  9.944651e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  2.096709e-06 -1.286239e-06  5.664611e-06 -5.295978e-05 -0.0002194325
#>  [2,] -2.218358e-04 -2.198194e-05 -3.099700e-04 -4.858837e-04 -0.0010397734
#>  [3,]  7.661277e-07 -1.322061e-06 -2.143148e-06 -2.858027e-05 -0.0001253966
#>  [4,]  9.625381e-05  7.971082e-06  1.343572e-04  1.517716e-04  0.0002254273
#>  [5,]  1.682476e-04  1.585260e-05  2.450031e-04  2.969696e-04  0.0004961131
#>  [6,]  1.843994e-04  1.536559e-05  2.686773e-04  3.748366e-04  0.0006822082
#>  [7,] -2.523555e-05 -4.247253e-06 -3.835691e-05 -6.727045e-05 -0.0001904516
#>  [8,]  4.936418e-04  4.043688e-05  6.787486e-04  9.285946e-04  0.0017058283
#>  [9,]  4.221615e-04  3.795441e-05  5.893703e-04  7.642436e-04  0.0013601229
#> [10,]  1.357908e-04  1.009170e-05  1.907373e-04  2.167311e-04  0.0002981866
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] -0.03051738 -0.03971708 -0.05388128 -0.05415229 -0.05058859 -0.003969194
#> [2,]  0.03051738  0.03971708  0.05388128  0.05415229  0.05058859  0.003969194
#>             [,7]       [,8]        [,9]       [,10]
#> [1,] -0.05917908 -0.0273879 -0.04039776 -0.02384497
#> [2,]  0.05917908  0.0273879  0.04039776  0.02384497
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.35808269 0.39831776 0.52015821 0.47977865 0.38232561 0.02470678
#>  [7] 0.78393332 0.40022913 0.43618508 0.30372893
#> 
#> $B[[2]]
#> [1]  0.1083709 -0.1654388
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  1.699182e-04 -1.348556e-03  1.285583e-04  7.210951e-04  1.204304e-03
#>  [6]  9.937546e-04 -6.839743e-05  3.285853e-03  2.994553e-03  1.035306e-03
#> 
#> $vB[[2]]
#> [1] -0.06957678  0.06957678
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#>  [2,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#>  [3,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#>  [4,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [5,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#>  [6,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#>  [7,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#>  [8,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [9,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [10,] 0.0323 0.2154 0.3085 0.3425 0.2990 0.1402 0.1235 0.1534 0.1901 0.2429
#> [11,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [12,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [13,] 0.0331 0.0193 0.0897 0.1734 0.1936 0.2803 0.3313 0.5020 0.6360 0.7096
#> [14,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [15,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [16,] 0.0335 0.4372 0.5533 0.5771 0.7022 0.7067 0.7367 0.7391 0.8622 0.9458
#> [17,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [18,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [19,] 0.0235 0.3199 0.2946 0.2484 0.2510 0.1806 0.1413 0.3019 0.3635 0.3887
#> [20,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#> [21,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [22,] 0.0203 0.1747 0.2198 0.2721 0.2105 0.1727 0.2040 0.1786 0.1318 0.2260
#> [23,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [24,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [25,] 0.0392 0.2471 0.2680 0.3049 0.2863 0.2294 0.1165 0.2127 0.2062 0.2222
#> [26,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [27,] 0.0109 0.1036 0.0972 0.0501 0.1546 0.3404 0.4804 0.6570 0.7738 0.7827
#> [28,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [29,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [30,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [31,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [32,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#> [33,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [34,] 0.0270 0.1520 0.1675 0.1370 0.1361 0.1345 0.2144 0.5354 0.6830 0.5600
#> [35,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [36,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [37,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [38,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [39,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [40,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#>  [2,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#>  [3,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#>  [4,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [5,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#>  [6,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#>  [7,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#>  [8,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [9,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [10,] 0.2120 0.0101 0.2395 0.3272 0.5949 0.8302 0.9045 0.9888 0.9912 0.9448
#> [11,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [12,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [13,] 0.8333 0.0423 0.8730 0.8073 0.7507 0.7526 0.7298 0.6177 0.4946 0.4531
#> [14,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [15,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [16,] 0.8782 0.0134 0.7913 0.5760 0.3061 0.0563 0.0239 0.2554 0.4862 0.5027
#> [17,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [18,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [19,] 0.2980 0.0220 0.2219 0.1624 0.1343 0.2046 0.3791 0.5771 0.7545 0.8406
#> [20,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#> [21,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [22,] 0.2358 0.0121 0.3107 0.3906 0.3631 0.4809 0.6531 0.7812 0.8395 0.9180
#> [23,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [24,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [25,] 0.3241 0.0108 0.4330 0.5071 0.5944 0.7078 0.7641 0.8878 0.9711 0.9880
#> [26,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [27,] 0.8152 0.0093 0.8129 0.8297 0.8535 0.8870 0.8894 0.8980 0.9667 1.0000
#> [28,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [29,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [30,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [31,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [32,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#> [33,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [34,] 0.3093 0.0092 0.3226 0.4430 0.5573 0.5782 0.6173 0.8132 0.9819 0.9823
#> [35,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [36,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [37,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [38,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [39,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [40,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#>  [2,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#>  [3,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#>  [4,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [5,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#>  [6,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#>  [7,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#>  [8,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [9,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [10,] 1.0000 0.9092 0.0298 0.7412 0.7691 0.7117 0.5304 0.2131 0.0928 0.1297
#> [11,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [12,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [13,] 0.4099 0.4540 0.0474 0.4124 0.3139 0.3194 0.3692 0.3776 0.4469 0.4777
#> [14,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [15,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [16,] 0.4402 0.2847 0.0696 0.1797 0.3560 0.3522 0.3321 0.3112 0.3638 0.0754
#> [17,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [18,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [19,] 0.8547 0.9036 0.0167 1.0000 0.9646 0.7912 0.6412 0.5986 0.6835 0.7771
#> [20,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#> [21,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [22,] 0.9769 0.8937 0.0380 0.7022 0.6500 0.5069 0.3903 0.3009 0.1565 0.0985
#> [23,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [24,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [25,] 0.9812 0.9464 0.0267 0.8542 0.6457 0.3397 0.3828 0.3204 0.1331 0.0440
#> [26,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [27,] 0.9134 0.6762 0.0121 0.4659 0.2895 0.2959 0.1746 0.2112 0.2569 0.2276
#> [28,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [29,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [30,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [31,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [32,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#> [33,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [34,] 0.9166 0.7423 0.0145 0.7736 0.8473 0.7352 0.6671 0.6083 0.6239 0.5972
#> [35,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [36,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [37,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [38,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [39,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [40,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#>  [2,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#>  [3,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#>  [4,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [5,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#>  [6,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#>  [7,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#>  [8,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [9,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [10,] 0.1159 0.1226 0.1768 0.0564 0.0345 0.1562 0.0824 0.1149 0.1694 0.0954
#> [11,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [12,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [13,] 0.4716 0.4664 0.3893 0.0818 0.4255 0.4064 0.3712 0.3863 0.2802 0.1283
#> [14,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [15,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [16,] 0.1834 0.1820 0.1815 0.1180 0.1593 0.0576 0.0954 0.1086 0.0812 0.0784
#> [17,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [18,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [19,] 0.8084 0.7426 0.6295 0.0516 0.5708 0.4433 0.3361 0.3795 0.4950 0.4373
#> [20,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#> [21,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [22,] 0.2200 0.2243 0.2736 0.0128 0.2152 0.2438 0.3154 0.2112 0.0991 0.0594
#> [23,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [24,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [25,] 0.1234 0.2030 0.1652 0.0257 0.1043 0.1066 0.2110 0.2417 0.1631 0.0769
#> [26,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [27,] 0.2149 0.1601 0.0371 0.0378 0.0117 0.0488 0.0288 0.0597 0.0431 0.0369
#> [28,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [29,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [30,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [31,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [32,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#> [33,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [34,] 0.5715 0.5242 0.2924 0.0278 0.1536 0.2003 0.2031 0.2207 0.1778 0.1353
#> [35,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [36,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [37,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [38,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [39,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [40,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#>  [2,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#>  [3,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#>  [4,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [5,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#>  [6,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#>  [7,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#>  [8,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [9,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [10,] 0.0080 0.0790 0.1255 0.0647 0.0760 0.0179 0.0051 0.0061 0.0093 0.0135
#> [11,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [12,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [13,] 0.1117 0.1303 0.0787 0.0436 0.0835 0.0224 0.0133 0.0078 0.0174 0.0176
#> [14,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [15,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [16,] 0.0487 0.0439 0.0586 0.0370 0.0348 0.0185 0.0302 0.0244 0.0232 0.0093
#> [17,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [18,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [19,] 0.2404 0.1128 0.1654 0.0933 0.0746 0.0225 0.0214 0.0221 0.0152 0.0083
#> [20,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#> [21,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [22,] 0.1940 0.1937 0.1082 0.0336 0.0537 0.0177 0.0209 0.0134 0.0094 0.0047
#> [23,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [24,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [25,] 0.0723 0.0912 0.0812 0.0496 0.0410 0.0101 0.0089 0.0083 0.0080 0.0026
#> [26,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [27,] 0.0025 0.0327 0.0257 0.0182 0.0679 0.0108 0.0124 0.0077 0.0023 0.0117
#> [28,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [29,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [30,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [31,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [32,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#> [33,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [34,] 0.1373 0.0749 0.0472 0.0325 0.0412 0.0179 0.0045 0.0084 0.0010 0.0018
#> [35,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [36,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [37,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [38,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [39,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [40,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#>  [2,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#>  [3,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#>  [4,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [5,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#>  [6,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#>  [7,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#>  [8,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [9,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [10,] 0.0063 0.0063 0.0034 0.0032 0.0062 0.0958 0.0067 0.0990 0.1018 0.1030
#> [11,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [12,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [13,] 0.0038 0.0129 0.0066 0.0044 0.0134 0.0756 0.0092 0.0374 0.0961 0.0548
#> [14,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [15,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [16,] 0.0159 0.0193 0.0032 0.0377 0.0126 0.1180 0.0156 0.1948 0.1607 0.3036
#> [17,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [18,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [19,] 0.0058 0.0023 0.0057 0.0052 0.0027 0.1121 0.0021 0.1258 0.1717 0.3074
#> [20,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#> [21,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [22,] 0.0045 0.0042 0.0028 0.0036 0.0013 0.0874 0.0016 0.1021 0.0852 0.1136
#> [23,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [24,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [25,] 0.0079 0.0042 0.0071 0.0044 0.0022 0.0491 0.0014 0.1053 0.1690 0.2105
#> [26,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [27,] 0.0053 0.0077 0.0076 0.0056 0.0055 0.0863 0.0039 0.1004 0.0664 0.0941
#> [28,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [29,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [30,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [31,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [32,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#> [33,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [34,] 0.0068 0.0039 0.0120 0.0132 0.0070 0.0757 0.0088 0.1026 0.1138 0.0794
#> [35,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [36,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [37,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [38,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [39,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [40,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5417311 0.6653420 0.8335890 0.8400318 0.7827110 0.2880087 0.8864679
#>  [2,] 0.5909019 0.6487501 0.7367744 0.7759859 0.7030666 0.3426793 0.8246668
#>  [3,] 0.5470247 0.7055472 0.8417864 0.8362002 0.7987225 0.3225139 0.8882785
#>  [4,] 0.5249810 0.5802717 0.7198080 0.7335896 0.6460102 0.2321674 0.7853958
#>  [5,] 0.6169639 0.7422162 0.8610737 0.8627060 0.8518470 0.3016702 0.9067340
#>  [6,] 0.5526376 0.6715998 0.8554709 0.8547670 0.8295305 0.2102849 0.8962663
#>  [7,] 0.5328925 0.6902696 0.8851580 0.8766727 0.8349370 0.2159566 0.9104135
#>  [8,] 0.5198901 0.6653081 0.8302230 0.8384251 0.7819196 0.2922316 0.8936041
#>  [9,] 0.6007986 0.7438865 0.8651888 0.8636998 0.8599971 0.2910178 0.9068020
#> [10,] 0.6119346 0.7542629 0.8729535 0.8739134 0.8759518 0.2908194 0.9101784
#> [11,] 0.5344374 0.6621855 0.8536399 0.8385921 0.7947863 0.2384689 0.8700856
#> [12,] 0.6262575 0.7158900 0.8281564 0.8003528 0.8018146 0.3554058 0.8578445
#> [13,] 0.5375072 0.6650956 0.8234863 0.8377351 0.7726766 0.2617090 0.8822153
#> [14,] 0.5645313 0.7214085 0.8735616 0.8743468 0.8424237 0.2450714 0.9124559
#> [15,] 0.6061913 0.7549326 0.8827933 0.8767147 0.8697352 0.2742649 0.9196621
#> [16,] 0.4783276 0.5051754 0.7502210 0.7387866 0.6517439 0.2439177 0.8084771
#> [17,] 0.4924106 0.5534311 0.6892012 0.7317386 0.6009200 0.1947496 0.8154819
#> [18,] 0.5718070 0.6321059 0.7805471 0.8113171 0.7598580 0.2070669 0.8627144
#> [19,] 0.5795335 0.6245587 0.7824500 0.8205904 0.7510492 0.2272082 0.8671696
#> [20,] 0.5837900 0.6663900 0.8144930 0.8300503 0.7962222 0.2120492 0.8780639
#> [21,] 0.5108564 0.6427563 0.8376910 0.8348647 0.7690741 0.2312961 0.8802881
#> [22,] 0.6164490 0.7312503 0.8430035 0.8406870 0.8323418 0.2855426 0.8882169
#> [23,] 0.5301917 0.5612402 0.7081612 0.7836678 0.6501771 0.2190162 0.8257156
#> [24,] 0.6172371 0.7742237 0.8783980 0.8799493 0.8784167 0.3256541 0.9144509
#> [25,] 0.6213756 0.7553714 0.8730383 0.8701555 0.8681979 0.2803352 0.9138585
#> [26,] 0.5614912 0.7185312 0.8889794 0.8739967 0.8440250 0.2917858 0.9153620
#> [27,] 0.6039189 0.7584636 0.9099189 0.9000739 0.8826766 0.3067759 0.9318692
#> [28,] 0.5986769 0.7445059 0.8779606 0.8685844 0.8656597 0.2608961 0.9162561
#> [29,] 0.5209222 0.6423327 0.8419677 0.8498843 0.7686629 0.2272100 0.8945142
#> [30,] 0.5831152 0.6868684 0.8549553 0.8503563 0.8134667 0.3305102 0.8830776
#> [31,] 0.5435862 0.6539642 0.8291526 0.8367678 0.8004010 0.1912399 0.8892595
#> [32,] 0.6316601 0.7520163 0.8509090 0.8740827 0.8510297 0.3113267 0.9158561
#> [33,] 0.4913648 0.6448795 0.8170565 0.8221700 0.7586295 0.2190905 0.8644101
#> [34,] 0.6170616 0.7253254 0.8719004 0.8821854 0.8558947 0.2739107 0.9186645
#> [35,] 0.6085369 0.7102065 0.8720480 0.8934980 0.8413992 0.2373745 0.9213203
#> [36,] 0.5971488 0.7105178 0.8374799 0.8331774 0.8231172 0.2697979 0.8842549
#> [37,] 0.6164840 0.7292285 0.8354313 0.8327532 0.8127778 0.3016520 0.8826298
#> [38,] 0.6250676 0.7526285 0.8723181 0.8667238 0.8607045 0.3133037 0.9081455
#> [39,] 0.6012484 0.7026219 0.8251478 0.8178465 0.8147695 0.2633642 0.8739781
#> [40,] 0.5349003 0.6738924 0.8445564 0.8415726 0.7913381 0.2948994 0.8831040
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5169715 0.7096272 0.5501161
#>  [2,] 0.5387306 0.6924622 0.5488252
#>  [3,] 0.5597442 0.7417067 0.5659815
#>  [4,] 0.4320975 0.5893935 0.4340390
#>  [5,] 0.6236592 0.7757312 0.5805921
#>  [6,] 0.5339904 0.7062896 0.4978981
#>  [7,] 0.5183795 0.7509482 0.5248150
#>  [8,] 0.5078515 0.7470664 0.5521768
#>  [9,] 0.6199600 0.7734684 0.5693481
#> [10,] 0.6488041 0.7948069 0.5664837
#> [11,] 0.4907177 0.6973980 0.4963761
#> [12,] 0.6282602 0.7284300 0.5727356
#> [13,] 0.4896905 0.7039011 0.5272498
#> [14,] 0.5460621 0.7526075 0.5690230
#> [15,] 0.6259152 0.7841602 0.5799365
#> [16,] 0.4444780 0.6056718 0.4448580
#> [17,] 0.3586647 0.5509237 0.4323704
#> [18,] 0.4757384 0.6796837 0.4388621
#> [19,] 0.4688391 0.6864132 0.4648451
#> [20,] 0.5031107 0.7073498 0.4660647
#> [21,] 0.4648152 0.6900814 0.4972416
#> [22,] 0.6180349 0.7563120 0.5609645
#> [23,] 0.3900450 0.6177434 0.4376151
#> [24,] 0.6578923 0.8034518 0.6175672
#> [25,] 0.6336349 0.7889478 0.5794463
#> [26,] 0.5876057 0.7685102 0.5847430
#> [27,] 0.6319577 0.8144005 0.6345471
#> [28,] 0.6173875 0.7750978 0.5635758
#> [29,] 0.4482595 0.7041755 0.4749607
#> [30,] 0.5973713 0.7593974 0.5646328
#> [31,] 0.5018084 0.6897432 0.4716631
#> [32,] 0.6032921 0.8021436 0.6018672
#> [33,] 0.4475335 0.6567724 0.4954371
#> [34,] 0.5733709 0.7904388 0.5794128
#> [35,] 0.5357678 0.7878097 0.5614747
#> [36,] 0.5978228 0.7498169 0.5505733
#> [37,] 0.5666130 0.7450182 0.5532398
#> [38,] 0.6419070 0.7764998 0.6093589
#> [39,] 0.5952793 0.7366095 0.5304088
#> [40,] 0.5291302 0.7376136 0.5518824
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4917416 0.5082584
#>  [2,] 0.4901971 0.5098029
#>  [3,] 0.4887459 0.5112541
#>  [4,] 0.5006418 0.4993582
#>  [5,] 0.4859391 0.5140609
#>  [6,] 0.4933012 0.5066988
#>  [7,] 0.4924583 0.5075417
#>  [8,] 0.4906379 0.5093621
#>  [9,] 0.4864220 0.5135780
#> [10,] 0.4846119 0.5153881
#> [11,] 0.4946518 0.5053482
#> [12,] 0.4858939 0.5141061
#> [13,] 0.4939714 0.5060286
#> [14,] 0.4909335 0.5090665
#> [15,] 0.4864642 0.5135358
#> [16,] 0.4973297 0.5026703
#> [17,] 0.5059979 0.4940021
#> [18,] 0.4967773 0.5032227
#> [19,] 0.4960262 0.5039738
#> [20,] 0.4949511 0.5050489
#> [21,] 0.4959728 0.5040272
#> [22,] 0.4872954 0.5127046
#> [23,] 0.5011453 0.4988547
#> [24,] 0.4831264 0.5168736
#> [25,] 0.4858446 0.5141554
#> [26,] 0.4873384 0.5126616
#> [27,] 0.4838409 0.5161591
#> [28,] 0.4873734 0.5126266
#> [29,] 0.4965493 0.5034507
#> [30,] 0.4857523 0.5142477
#> [31,] 0.4957513 0.5042487
#> [32,] 0.4856720 0.5143280
#> [33,] 0.4981277 0.5018723
#> [34,] 0.4877441 0.5122559
#> [35,] 0.4902102 0.5097898
#> [36,] 0.4883442 0.5116558
#> [37,] 0.4896081 0.5103919
#> [38,] 0.4848466 0.5151534
#> [39,] 0.4891156 0.5108844
#> [40,] 0.4901348 0.5098652
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]       [,2]      [,3]     [,4]      [,5]       [,6]     [,7]
#>  [1,]  0.16731359 0.68719195 1.6112801 1.658464 1.2815351 -0.9050750 2.055159
#>  [2,]  0.36769512 0.61354980 1.0292704 1.242425 0.8619436 -0.6513771 1.548291
#>  [3,]  0.18865632 0.87385474 1.6715801 1.630223 1.3783290 -0.7422429 2.073276
#>  [4,]  0.10000726 0.32388891 0.9435094 1.012912 0.6015471 -1.1961131 1.297392
#>  [5,]  0.47668109 1.05751918 1.8242363 1.837950 1.7491615 -0.8393571 2.274394
#>  [6,]  0.21133334 0.71542943 1.7781713 1.772490 1.5823034 -1.3232087 2.156410
#>  [7,]  0.13176044 0.80138002 2.0422091 1.961292 1.6210294 -1.2893866 2.318694
#>  [8,]  0.07960247 0.68703987 1.5872088 1.646556 1.2768884 -0.8845704 2.128096
#>  [9,]  0.40879377 1.06626767 1.8590726 1.846365 1.8152659 -0.8904461 2.275198
#> [10,]  0.45545166 1.12147883 1.9273295 1.936012 1.9546410 -0.8914076 2.315816
#> [11,]  0.13796793 0.67304890 1.7634397 1.647790 1.3540216 -1.1610921 1.901716
#> [12,]  0.51619464 0.92416500 1.5726171 1.388501 1.3976747 -0.5953609 1.797501
#> [13,]  0.15031106 0.68608570 1.5401487 1.641472 1.2234868 -1.0371049 2.013577
#> [14,]  0.25957310 0.95145912 1.9328238 1.939952 1.6763733 -1.1250737 2.343997
#> [15,]  0.43133045 1.12509543 2.0191523 1.961680 1.8986195 -0.9730907 2.437765
#> [16,] -0.08674387 0.02070242 1.0997914 1.039671 0.6267137 -1.1313193 1.440145
#> [17,] -0.03035999 0.21454365 0.7963877 1.003461 0.4093000 -1.4194386 1.486032
#> [18,]  0.28922752 0.54126161 1.2688572 1.458591 1.1519009 -1.3426970 1.838020
#> [19,]  0.32085863 0.50894307 1.2800014 1.520352 1.1042160 -1.2241429 1.876162
#> [20,]  0.33835146 0.69190224 1.4794735 1.585983 1.3628478 -1.3126172 1.974222
#> [21,]  0.04343229 0.58734727 1.6411472 1.620504 1.2030906 -1.2010069 1.995161
#> [22,]  0.47450300 1.00097550 1.6807479 1.663349 1.6023161 -0.9171321 2.072656
#> [23,]  0.12091402 0.24619674 0.8864701 1.287170 0.6198180 -1.2714088 1.555562
#> [24,]  0.47783729 1.23231607 1.9773465 1.991950 1.9775212 -0.7279075 2.369234
#> [25,]  0.49539098 1.12746825 1.9280941 1.902335 1.8851178 -0.9427997 2.361685
#> [26,]  0.24721612 0.93718773 2.0803585 1.936768 1.6884867 -0.8867266 2.380936
#> [27,]  0.42182076 1.14427453 2.3126450 2.198046 2.0180245 -0.8152358 2.615763
#> [28,]  0.39995534 1.06952124 1.9732579 1.888500 1.8631158 -1.0413161 2.392532
#> [29,]  0.08373787 0.58550320 1.6729424 1.733694 1.2007765 -1.2241325 2.137704
#> [30,]  0.33557495 0.78551932 1.7740075 1.737398 1.4726954 -0.7058786 2.021903
#> [31,]  0.17478836 0.63651051 1.5796337 1.634373 1.3888028 -1.4419735 2.083199
#> [32,]  0.53934527 1.10939500 1.7417487 1.937549 1.7427002 -0.7939242 2.387330
#> [33,] -0.03454435 0.59660609 1.4965307 1.531119 1.1451806 -1.2709744 1.852412
#> [34,]  0.47709481 0.97103351 1.9178674 2.013290 1.7816032 -0.9748709 2.424338
#> [35,]  0.44116626 0.89638700 1.9191892 2.126980 1.6686760 -1.1671280 2.460423
#> [36,]  0.39359922 0.89790029 1.6395953 1.608316 1.5376113 -0.9956483 2.033355
#> [37,]  0.47465117 0.99071189 1.6246198 1.605267 1.4681618 -0.8394435 2.017573
#> [38,]  0.51111397 1.11268038 1.9216117 1.872296 1.8211534 -0.7847189 2.291198
#> [39,]  0.41066968 0.85981437 1.5516212 1.501825 1.4813044 -1.0285555 1.936599
#> [40,]  0.13982872 0.72584315 1.6925283 1.669976 1.3330103 -0.8717061 2.022158
#>               [,8]      [,9]        [,10]
#>  [1,]  0.067911982 0.8935739  0.201139763
#>  [2,]  0.155233250 0.8116559  0.195924989
#>  [3,]  0.240123830 1.0548584  0.265474427
#>  [4,] -0.273298511 0.3614589 -0.265390710
#>  [5,]  0.505108784 1.2409606  0.325204504
#>  [6,]  0.136171763 0.8774313 -0.008407755
#>  [7,]  0.073551317 1.1036759  0.099341776
#>  [8,]  0.031408605 1.0830272  0.209469747
#>  [9,]  0.489378365 1.2280005  0.279191793
#> [10,]  0.613786831 1.3541474  0.267519058
#> [11,] -0.037133403 0.8349380 -0.014495689
#> [12,]  0.524760127 0.9866714  0.293021116
#> [13,] -0.041244038 0.8659442  0.109107284
#> [14,]  0.184772430 1.1125676  0.277866041
#> [15,]  0.514732553 1.2900768  0.322512707
#> [16,] -0.223007592 0.4291549 -0.221468876
#> [17,] -0.581164366 0.2044036 -0.272186281
#> [18,] -0.097122549 0.7523185 -0.245781348
#> [19,] -0.124805261 0.7834034 -0.140851858
#> [20,]  0.012442809 0.8825473 -0.135950181
#> [21,] -0.140972312 0.8004999 -0.011033524
#> [22,]  0.481215612 1.1325652  0.245077370
#> [23,] -0.447122952 0.4799811 -0.250846922
#> [24,]  0.653915598 1.4080096  0.479234831
#> [25,]  0.547842382 1.3185945  0.320500662
#> [26,]  0.354076341 1.1999179  0.342274883
#> [27,]  0.540624604 1.4788612  0.551774263
#> [28,]  0.478473980 1.2373236  0.255687016
#> [29,] -0.207705573 0.8672612 -0.100241001
#> [30,]  0.394524040 1.1493785  0.259985810
#> [31,]  0.007233485 0.7989189 -0.113469370
#> [32,]  0.419201343 1.3997461  0.413251185
#> [33,] -0.210641379 0.6489439 -0.018252091
#> [34,]  0.295617708 1.3275723  0.320363336
#> [35,]  0.143315852 1.3117728  0.247149162
#> [36,]  0.396401742 1.0976361  0.202987482
#> [37,]  0.268045584 1.0722163  0.213769420
#> [38,]  0.583650719 1.2453842  0.444618295
#> [39,]  0.385833437 1.0284200  0.121785309
#> [40,]  0.116652898 1.0336019  0.208279159
#> 
#> $pre[[3]]
#>               [,1]          [,2]
#>  [1,] -0.021525858  1.151087e-02
#>  [2,] -0.024008319  1.520826e-02
#>  [3,] -0.024167750  2.085638e-02
#>  [4,] -0.012318160 -1.488525e-02
#>  [5,] -0.031735453  2.452310e-02
#>  [6,] -0.022348408  4.448393e-03
#>  [7,] -0.023598263  6.570760e-03
#>  [8,] -0.021781382  1.567155e-02
#>  [9,] -0.030427211  2.389808e-02
#> [10,] -0.033414668  2.815715e-02
#> [11,] -0.018938463  2.455345e-03
#> [12,] -0.030271416  2.616806e-02
#> [13,] -0.019172714  4.943002e-03
#> [14,] -0.025662366  1.060765e-02
#> [15,] -0.032275665  2.188072e-02
#> [16,] -0.015712797 -5.031436e-03
#> [17,] -0.005584435 -2.957716e-02
#> [18,] -0.017212362 -4.321416e-03
#> [19,] -0.018261527 -2.366184e-03
#> [20,] -0.020225665 -2.932829e-05
#> [21,] -0.016964174 -8.551339e-04
#> [22,] -0.031043891  1.978548e-02
#> [23,] -0.010827175 -1.540829e-02
#> [24,] -0.034733588  3.278627e-02
#> [25,] -0.033282786  2.335413e-02
#> [26,] -0.029076014  2.158128e-02
#> [27,] -0.035385621  2.927328e-02
#> [28,] -0.031129506  1.938772e-02
#> [29,] -0.015903547 -2.100660e-03
#> [30,] -0.029936511  2.706976e-02
#> [31,] -0.019316293 -2.321219e-03
#> [32,] -0.031725015  2.560263e-02
#> [33,] -0.013445965 -5.956578e-03
#> [34,] -0.030220293  1.881308e-02
#> [35,] -0.028717350  1.044690e-02
#> [36,] -0.029662032  1.696979e-02
#> [37,] -0.026006506  1.556718e-02
#> [38,] -0.033824329  2.680779e-02
#> [39,] -0.028710944  1.483365e-02
#> [40,] -0.023430726  1.603514e-02
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5082584 -0.5082584
#>  [2,] -0.4901971  0.4901971
#>  [3,] -0.4887459  0.4887459
#>  [4,]  0.4993582 -0.4993582
#>  [5,] -0.4859391  0.4859391
#>  [6,]  0.5066988 -0.5066988
#>  [7,]  0.5075417 -0.5075417
#>  [8,]  0.5093621 -0.5093621
#>  [9,]  0.5135780 -0.5135780
#> [10,]  0.5153881 -0.5153881
#> [11,] -0.4946518  0.4946518
#> [12,] -0.4858939  0.4858939
#> [13,]  0.5060286 -0.5060286
#> [14,]  0.5090665 -0.5090665
#> [15,]  0.5135358 -0.5135358
#> [16,]  0.5026703 -0.5026703
#> [17,]  0.4940021 -0.4940021
#> [18,]  0.5032227 -0.5032227
#> [19,]  0.5039738 -0.5039738
#> [20,]  0.5050489 -0.5050489
#> [21,]  0.5040272 -0.5040272
#> [22,]  0.5127046 -0.5127046
#> [23,] -0.5011453  0.5011453
#> [24,] -0.4831264  0.4831264
#> [25,]  0.5141554 -0.5141554
#> [26,] -0.4873384  0.4873384
#> [27,] -0.4838409  0.4838409
#> [28,]  0.5126266 -0.5126266
#> [29,] -0.4965493  0.4965493
#> [30,] -0.4857523  0.4857523
#> [31,]  0.5042487 -0.5042487
#> [32,] -0.4856720  0.4856720
#> [33,]  0.5018723 -0.5018723
#> [34,] -0.4877441  0.4877441
#> [35,] -0.4902102  0.4902102
#> [36,]  0.5116558 -0.5116558
#> [37,] -0.4896081  0.4896081
#> [38,] -0.4848466  0.4848466
#> [39,]  0.5108844 -0.5108844
#> [40,] -0.4901348  0.4901348
#> 
#> $L
#> [1] 0.6886038 0.7004215 0.7244254 0.7206250 0.6923630 0.6924290
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609