Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11          V12         V13         V14
#>  [1,] -0.3290943 -0.117577082 -0.089944653 -0.055356801  0.01014745 -0.01701301
#>  [2,] -0.3798239 -0.207418981 -0.101829264 -0.124768296 -0.13337626 -0.12161467
#>  [3,] -0.2358729 -0.080780492 -0.027888176 -0.153009801 -0.06189978 -0.03695649
#>  [4,] -0.2383647 -0.254565582 -0.189552368 -0.158518809 -0.16357944 -0.10088617
#>  [5,] -0.2709176 -0.007751705 -0.070437411 -0.014157141  0.03259138 -0.03064354
#>  [6,] -0.2462340 -0.081251108 -0.112140781 -0.129206255 -0.05675220  0.04725774
#>  [7,] -0.3777186 -0.037448976 -0.043829666 -0.142853305 -0.02975408 -0.07184167
#>  [8,] -0.2758894 -0.176998420 -0.210231919 -0.131249351 -0.17425493 -0.16986089
#>  [9,] -0.2548885 -0.056550418 -0.129442360 -0.058725275 -0.09875783 -0.05366162
#> [10,] -0.2491921 -0.096548500 -0.005649727  0.004301128 -0.08431186 -0.03287694
#>                V15         V16          V17          V18          V19
#>  [1,] -0.073426975 -0.04465134  0.019772286 -0.004580380  0.067895504
#>  [2,] -0.104846087 -0.01654745 -0.095861199 -0.036431663  0.054036785
#>  [3,] -0.005019433 -0.12247541  0.029501621 -0.030274043 -0.005530525
#>  [4,] -0.122101674 -0.09594499  0.006139819  0.045964752 -0.114084891
#>  [5,]  0.015358761 -0.02623808 -0.017565681 -0.016528724  0.006973106
#>  [6,]  0.051651432 -0.08958175 -0.005222416  0.124063555  0.062152579
#>  [7,]  0.003181513  0.01250646  0.038934796 -0.020485362  0.062032083
#>  [8,] -0.123218133 -0.05813375 -0.016806368 -0.053623923 -0.037261283
#>  [9,] -0.045426904 -0.01538792 -0.048741039  0.089854951  0.147843324
#> [10,] -0.024369320  0.03074311  0.058141839 -0.005840616 -0.009720737
#>               V2           V20         V21         V22          V23         V24
#>  [1,] -0.3029728  1.130812e-01  0.23125453 0.166824995  0.330064523  0.22625590
#>  [2,] -0.2708751  7.585795e-02 -0.06490374 0.006088164 -0.029920826  0.08652720
#>  [3,] -0.2942763  5.827925e-05  0.03192472 0.063134277  0.019404113  0.11689256
#>  [4,] -0.3114071 -5.236482e-02 -0.02180290 0.094089046  0.001172140 -0.01790695
#>  [5,] -0.2379695  1.139476e-01  0.20422640 0.090562004  0.115216698 -0.02006352
#>  [6,] -0.2106113  3.847434e-02  0.02294459 0.072970597  0.133519833  0.19819746
#>  [7,] -0.3218722  1.224675e-02  0.07298165 0.127736499  0.008776613  0.12677179
#>  [8,] -0.2497255 -5.625111e-02  0.06150634 0.056285450  0.025998715 -0.09005692
#>  [9,] -0.2648044  1.928998e-01  0.14886208 0.050149828  0.195186014  0.11153359
#> [10,] -0.2509283  6.410026e-02 -0.00447797 0.112537237  0.076500023  0.15459497
#>              V25         V26        V27         V28          V29         V3
#>  [1,] 0.17859529 0.280587510 0.16085360  0.14712610  0.299556093 -0.2280627
#>  [2,] 0.10197944 0.123341340 0.13708286  0.20024304  0.138195657 -0.2936903
#>  [3,] 0.16296369 0.170098761 0.16910779  0.15219727  0.076962148 -0.2096482
#>  [4,] 0.04835893 0.051328955 0.20443231  0.06202685  0.053185318 -0.1583803
#>  [5,] 0.04749337 0.003140217 0.08123687 -0.03125160 -0.080845418 -0.2224276
#>  [6,] 0.21492852 0.072989408 0.21095291  0.22261316 -0.003951523 -0.2897264
#>  [7,] 0.03374686 0.165436868 0.08977875  0.06433452  0.126711858 -0.3939366
#>  [8,] 0.10215450 0.041992024 0.15633750  0.09917708 -0.066522141 -0.3231210
#>  [9,] 0.22875567 0.224670859 0.10011013  0.07941993  0.164181207 -0.3214696
#> [10,] 0.12658936 0.101285145 0.17822972  0.16983002  0.134168728 -0.3430365
#>               V30          V31          V32         V33          V34
#>  [1,]  0.17397700  0.222931252  0.170811628  0.14052805  0.058361115
#>  [2,]  0.04784217 -0.032794617 -0.042888823  0.06250695 -0.038697008
#>  [3,]  0.05468882  0.001438407 -0.094824958 -0.12194956 -0.039084398
#>  [4,] -0.08564775 -0.118132340 -0.018066421 -0.18974685 -0.114435750
#>  [5,] -0.13056878 -0.132324187 -0.006453756 -0.17380178  0.004970425
#>  [6,]  0.10487632 -0.079231812 -0.071637984 -0.11351389 -0.104840987
#>  [7,]  0.17850611  0.149672402  0.045726377  0.05144218  0.078810136
#>  [8,]  0.02654004  0.039951767 -0.050618258 -0.14154150 -0.018167003
#>  [9,]  0.10542478  0.052675121 -0.012915724  0.06434047 -0.079728197
#> [10,]  0.02239590 -0.024231137  0.051938644  0.10787040 -0.048132334
#>               V35          V36          V37          V38         V39         V4
#>  [1,] -0.02735536  0.098280760 -0.037783555 -0.012781187  0.04615305 -0.2965284
#>  [2,]  0.02838916  0.007502441 -0.074968178 -0.147630209  0.04691234 -0.3352275
#>  [3,] -0.15931987 -0.062156760 -0.015827883 -0.058543253 -0.17083283 -0.2748625
#>  [4,] -0.21162256 -0.184942662  0.001290999 -0.095603048 -0.03197354 -0.2851834
#>  [5,] -0.13365307 -0.014808666 -0.078257695 -0.002255116 -0.04628258 -0.3097405
#>  [6,] -0.08146523 -0.142640890 -0.001202472 -0.040102813 -0.13768489 -0.2606738
#>  [7,]  0.13096179  0.002593535 -0.080285901 -0.065493422 -0.01674282 -0.3732169
#>  [8,] -0.12673179 -0.126794296 -0.083099442 -0.035962315 -0.06321093 -0.3366875
#>  [9,] -0.07708083  0.002381590 -0.009425175 -0.036246734 -0.08520989 -0.1957191
#> [10,]  0.05586713 -0.079857051 -0.035629198 -0.048644636 -0.05106937 -0.1361005
#>                V40          V41         V42          V43          V44
#>  [1,] -0.077981254  0.005933819  0.03921447 -0.101686635  0.028709872
#>  [2,] -0.038260935 -0.010839333 -0.05251359 -0.066615632 -0.076600299
#>  [3,] -0.034234742 -0.076230944 -0.23815499 -0.220648905 -0.276692925
#>  [4,] -0.190333284 -0.146858104 -0.08343837 -0.114379307 -0.071768046
#>  [5,] -0.005860047 -0.123069794 -0.11067268 -0.128363144 -0.192434361
#>  [6,] -0.081016190 -0.078372452 -0.09346769  0.004064168 -0.199353590
#>  [7,]  0.022245455 -0.048068058 -0.11789265  0.029961685 -0.035996444
#>  [8,] -0.064681530 -0.251368009 -0.11971144 -0.055117617 -0.113466340
#>  [9,] -0.093186662 -0.029061752 -0.05306091 -0.147933006 -0.105523553
#> [10,]  0.053102012  0.051044432  0.04719306  0.031652911  0.001032864
#>                V45          V46        V47        V48        V49         V5
#>  [1,] -0.102555459 -0.005681656 -0.1649962 -0.2348826 -0.3319604 -0.3081140
#>  [2,] -0.188548257 -0.041764878 -0.1766804 -0.1761925 -0.3041400 -0.1994052
#>  [3,] -0.265808938 -0.162771092 -0.1337231 -0.3008989 -0.2876355 -0.2599753
#>  [4,] -0.100046228 -0.238523814 -0.1737451 -0.1376580 -0.3113703 -0.1458669
#>  [5,] -0.136590123 -0.231131541 -0.2599720 -0.2886746 -0.2400380 -0.2354391
#>  [6,] -0.132103894 -0.109647593 -0.1217196 -0.1492826 -0.2517273 -0.1439376
#>  [7,] -0.065663477 -0.149821043 -0.1859567 -0.1718918 -0.2901555 -0.3240778
#>  [8,] -0.134494728 -0.223318833 -0.2614116 -0.2024953 -0.1884770 -0.1339657
#>  [9,] -0.083464004 -0.203538630 -0.1123674 -0.1554729 -0.2245645 -0.2908081
#> [10,]  0.001905388 -0.116244946 -0.1928700 -0.1837185 -0.3149466 -0.2416459
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.3613735 -0.3643707 -0.3410891 -0.3074771 -0.4104814 -0.2626696
#>  [2,] -0.3089188 -0.3079921 -0.2770226 -0.2380613 -0.3187928 -0.3376743
#>  [3,] -0.3395812 -0.3168254 -0.2314760 -0.3680545 -0.2947010 -0.3769562
#>  [4,] -0.2718730 -0.2678421 -0.2176083 -0.1970939 -0.2197507 -0.2138360
#>  [5,] -0.3698570 -0.3571054 -0.3165033 -0.3423028 -0.2584941 -0.2366930
#>  [6,] -0.2726626 -0.2361405 -0.3736269 -0.3838078 -0.2399917 -0.2839042
#>  [7,] -0.3095814 -0.3024640 -0.4103162 -0.3302972 -0.3506076 -0.4241181
#>  [8,] -0.3748910 -0.3926500 -0.3680364 -0.3425826 -0.3950640 -0.2380193
#>  [9,] -0.2884457 -0.4383088 -0.3619150 -0.3954750 -0.4360576 -0.3548560
#> [10,] -0.3150091 -0.3351636 -0.3480869 -0.3612351 -0.3728881 -0.3652698
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3790356 -0.4113873 -0.3429312 -0.3542168 -0.1770089 -0.4294611
#>  [2,] -0.4293469 -0.3816331 -0.3422965 -0.2678353 -0.2476748 -0.2713308
#>  [3,] -0.3884897 -0.3518517 -0.3745289 -0.3964359 -0.2485609 -0.3143974
#>  [4,] -0.3037743 -0.2751047 -0.2256888 -0.2891872 -0.1483631 -0.2486062
#>  [5,] -0.3743805 -0.2542422 -0.2237190 -0.2723273 -0.2243495 -0.3607285
#>  [6,] -0.2423611 -0.2448361 -0.3848479 -0.3461872 -0.2617920 -0.3454035
#>  [7,] -0.3621609 -0.3919054 -0.4416832 -0.3035297 -0.1114834 -0.4515696
#>  [8,] -0.2852016 -0.3818552 -0.2781170 -0.4029065 -0.2679097 -0.2424818
#>  [9,] -0.2746135 -0.3249606 -0.3227596 -0.4071471 -0.2733621 -0.2563425
#> [10,] -0.2829269 -0.3985799 -0.3534350 -0.2884683 -0.2299504 -0.3859652
#>               V7          V8         V9
#>  [1,] -0.1600229 -0.21748279 -0.0838615
#>  [2,] -0.2124401 -0.17138839 -0.1517436
#>  [3,] -0.1157228 -0.25780010 -0.1598033
#>  [4,] -0.1910128 -0.11853500 -0.2332824
#>  [5,] -0.1966033 -0.07751217 -0.1409122
#>  [6,] -0.1987579 -0.25608234 -0.1468923
#>  [7,] -0.1096089 -0.19691643 -0.2053122
#>  [8,] -0.1729083 -0.16020357 -0.0893989
#>  [9,] -0.1861180 -0.10439688 -0.0579140
#> [10,] -0.2217254 -0.20829217 -0.1806107
#> 
#> $W[[2]]
#>             [,1]       [,2]       [,3]        [,4]       [,5]       [,6]
#> [1,] -0.09049878 -0.1024559 -0.2083406 -0.07349519 -0.0231410 -0.1186443
#> [2,]  0.08430869  0.1928477  0.1227429  0.08627066  0.2161329  0.1692365
#>            [,7]        [,8]        [,9]      [,10]
#> [1,] -0.2627187 -0.06931453 -0.10925447 -0.2339667
#> [2,]  0.2093518  0.12612508  0.06758663  0.2099937
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10          V11          V12          V13
#>  [1,]  1.090158e-04  8.192389e-04 7.583952e-04 7.304181e-04 9.025311e-04
#>  [2,]  5.801232e-05  3.708952e-04 3.894753e-04 3.788420e-04 4.304630e-04
#>  [3,] -7.432282e-06 -7.597963e-07 6.115362e-05 9.531424e-05 2.390898e-05
#>  [4,]  5.111558e-05  2.562714e-04 2.281005e-04 2.042489e-04 3.541571e-04
#>  [5,]  3.136095e-05  2.336723e-04 2.766469e-04 2.744771e-04 2.540505e-04
#>  [6,]  1.109538e-04  7.153816e-04 7.150405e-04 6.730174e-04 7.724635e-04
#>  [7,]  4.233051e-05  2.770547e-04 3.925585e-04 3.715076e-04 2.289981e-04
#>  [8,]  2.529269e-05  1.239266e-04 1.169379e-04 1.231984e-04 2.051392e-04
#>  [9,]  1.815723e-04  1.190000e-03 1.079267e-03 1.036482e-03 1.343024e-03
#> [10,]  5.781711e-05  3.773439e-04 4.684164e-04 4.076133e-04 2.864365e-04
#>                 V14           V15           V16           V17           V18
#>  [1,]  1.237575e-03  0.0018235573  0.0022723572  0.0025062135  0.0024462578
#>  [2,]  4.957731e-04  0.0006086632  0.0007240110  0.0007907120  0.0008085580
#>  [3,] -1.501546e-04 -0.0003728021 -0.0005395088 -0.0005796319 -0.0005230167
#>  [4,]  5.867455e-04  0.0008496634  0.0011244103  0.0012227827  0.0011929518
#>  [5,]  1.950278e-04  0.0001400476  0.0001046982  0.0001098072  0.0001501383
#>  [6,]  9.402843e-04  0.0012291215  0.0014764177  0.0016050120  0.0016151510
#>  [7,] -7.162468e-05 -0.0004597872 -0.0007771755 -0.0009057004 -0.0008178141
#>  [8,]  3.144187e-04  0.0004381160  0.0005816372  0.0006430467  0.0006373178
#>  [9,]  1.919145e-03  0.0028622478  0.0036037956  0.0039795740  0.0039257922
#> [10,]  5.125824e-05 -0.0002495543 -0.0005195634 -0.0006495616 -0.0005915581
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0022733061  1.224002e-04  0.0022666659  2.255187e-03  2.661374e-03
#>  [2,]  0.0008764813  6.003642e-05  0.0009796157  1.021148e-03  1.113757e-03
#>  [3,] -0.0003984530 -9.644880e-06 -0.0003207794 -2.879712e-04 -4.227666e-04
#>  [4,]  0.0011412347  5.072729e-05  0.0011868338  1.209496e-03  1.441028e-03
#>  [5,]  0.0002746267  3.278314e-05  0.0003862147  4.364840e-04  4.023870e-04
#>  [6,]  0.0016922576  1.190269e-04  0.0018518224  1.906255e-03  2.099637e-03
#>  [7,] -0.0004181006  4.760378e-05 -0.0001250034  1.468810e-06 -3.254794e-04
#>  [8,]  0.0006081967  2.419020e-05  0.0006288711  6.375044e-04  7.666450e-04
#>  [9,]  0.0036584692  1.950485e-04  0.0036823690  3.661777e-03  4.345753e-03
#> [10,] -0.0002119273  6.322969e-05  0.0001012854  2.310047e-04 -7.754232e-05
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0031291684  0.0032702876  0.0035919496  0.0037442861  0.0042594873
#>  [2,]  0.0012052771  0.0012586037  0.0013493159  0.0013985547  0.0015407070
#>  [3,] -0.0006185878 -0.0006599731 -0.0007097438 -0.0007472695 -0.0008587453
#>  [4,]  0.0017381400  0.0018592903  0.0020391060  0.0021356148  0.0022931903
#>  [5,]  0.0003381985  0.0003443243  0.0003438635  0.0003542600  0.0003844479
#>  [6,]  0.0023256348  0.0024202726  0.0025858349  0.0026949256  0.0030166188
#>  [7,] -0.0007207267 -0.0007741694 -0.0009481866 -0.0009659378 -0.0011232550
#>  [8,]  0.0009200298  0.0009825916  0.0010932504  0.0011356738  0.0012079136
#>  [9,]  0.0051734541  0.0053848031  0.0058289366  0.0060605746  0.0068574270
#> [10,] -0.0004260125 -0.0004603263 -0.0005995018 -0.0005968449 -0.0006926773
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0043993098  0.0040233250  1.438757e-04  0.0038791761  0.0036492728
#>  [2,]  0.0016165022  0.0014944639  6.479554e-05  0.0013958207  0.0012036783
#>  [3,] -0.0008654211 -0.0008283978 -1.289081e-05 -0.0008402803 -0.0008767085
#>  [4,]  0.0023193981  0.0021785029  5.722557e-05  0.0020498673  0.0018757137
#>  [5,]  0.0004234997  0.0003676653  3.332352e-05  0.0003001919  0.0001715547
#>  [6,]  0.0031871857  0.0029737697  1.288690e-04  0.0028314920  0.0025179857
#>  [7,] -0.0010610770 -0.0010061556  3.373462e-05 -0.0010628487 -0.0012448868
#>  [8,]  0.0012272128  0.0011624097  2.884995e-05  0.0010919755  0.0009922280
#>  [9,]  0.0071601477  0.0067098906  2.292709e-04  0.0064942843  0.0060673770
#> [10,] -0.0006210984 -0.0006321835  4.787290e-05 -0.0006900336 -0.0008921953
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0027541563  0.0026577061  2.739642e-03  3.134386e-03  3.192978e-03
#>  [2,]  0.0009408508  0.0009114527  8.981933e-04  9.597144e-04  9.150462e-04
#>  [3,] -0.0006585845 -0.0006948554 -7.675267e-04 -9.109705e-04 -9.587904e-04
#>  [4,]  0.0014464892  0.0014681027  1.526524e-03  1.680212e-03  1.666926e-03
#>  [5,]  0.0001514272  0.0001198348  6.775907e-05  2.391456e-05 -2.671657e-05
#>  [6,]  0.0019774985  0.0019269508  1.939570e-03  2.110328e-03  2.047228e-03
#>  [7,] -0.0008536523 -0.0009038929 -1.047682e-03 -1.325389e-03 -1.484067e-03
#>  [8,]  0.0007629257  0.0007692566  8.037207e-04  8.765936e-04  8.627896e-04
#>  [9,]  0.0046575134  0.0046030501  4.805935e-03  5.418958e-03  5.446166e-03
#> [10,] -0.0005993957 -0.0006533583 -7.922035e-04 -9.961591e-04 -1.139478e-03
#>                 V37           V38           V39           V4           V40
#>  [1,]  2.856134e-03  0.0018973530  1.545234e-03 3.159364e-05  0.0018909689
#>  [2,]  8.451268e-04  0.0006639642  5.576128e-04 5.559583e-05  0.0005694584
#>  [3,] -7.827957e-04 -0.0003751644 -2.320857e-04 1.781102e-05 -0.0004016309
#>  [4,]  1.402676e-03  0.0008794804  6.414004e-04 2.392517e-05  0.0007663815
#>  [5,]  4.305487e-05  0.0001699074  2.066229e-04 4.778165e-05  0.0001228872
#>  [6,]  1.861714e-03  0.0013914951  1.161325e-03 1.011524e-04  0.0012514436
#>  [7,] -1.182094e-03 -0.0003798297 -6.174097e-05 1.236468e-04 -0.0004365996
#>  [8,]  7.199627e-04  0.0004557937  3.198159e-04 1.135343e-05  0.0003774015
#>  [9,]  4.766122e-03  0.0031283688  2.410012e-03 8.487197e-05  0.0029743751
#> [10,] -8.552427e-04 -0.0001698440  1.043155e-04 1.255334e-04 -0.0002305568
#>                 V41           V42           V43           V44          V45
#>  [1,]  0.0017800104  1.469367e-03  9.387331e-04  7.493828e-04 4.417763e-04
#>  [2,]  0.0005430536  5.197632e-04  4.142788e-04  3.529968e-04 2.963759e-04
#>  [3,] -0.0003859042 -2.304383e-04 -7.373702e-05 -2.248615e-05 7.334474e-05
#>  [4,]  0.0007372597  5.871357e-04  3.654869e-04  2.825091e-04 1.354126e-04
#>  [5,]  0.0001087726  1.875691e-04  2.233339e-04  2.053694e-04 2.343545e-04
#>  [6,]  0.0011890441  1.086316e-03  8.341915e-04  6.941941e-04 5.566340e-04
#>  [7,] -0.0004271594 -7.900186e-05  2.493112e-04  2.719589e-04 5.272903e-04
#>  [8,]  0.0003656476  2.904068e-04  1.723069e-04  1.357784e-04 5.208756e-05
#>  [9,]  0.0028153466  2.294429e-03  1.467876e-03  1.136783e-03 5.925183e-04
#> [10,] -0.0002323756  7.568839e-05  3.454494e-04  3.402147e-04 5.765145e-04
#>                V46          V47          V48          V49           V5
#>  [1,] 2.971491e-04 4.013932e-04 2.469199e-04 1.217852e-04 1.441764e-04
#>  [2,] 2.431606e-04 2.119998e-04 1.500814e-04 7.309276e-05 9.998465e-05
#>  [3,] 7.941541e-05 1.580186e-05 2.260362e-05 1.548357e-05 8.031136e-06
#>  [4,] 1.015357e-04 1.375825e-04 9.264353e-05 3.524608e-05 7.005935e-05
#>  [5,] 2.010728e-04 1.354161e-04 1.030135e-04 5.555524e-05 6.902819e-05
#>  [6,] 4.383544e-04 3.940245e-04 2.730635e-04 1.354054e-04 1.860117e-04
#>  [7,] 4.775771e-04 2.095503e-04 1.818102e-04 1.120759e-04 1.275727e-04
#>  [8,] 4.120011e-05 7.078998e-05 4.708011e-05 1.531719e-05 3.538434e-05
#>  [9,] 4.026042e-04 5.788467e-04 3.623910e-04 1.666472e-04 2.528592e-04
#> [10,] 5.147501e-04 2.479581e-04 2.049257e-04 1.240724e-04 1.376093e-04
#>                 V50          V51          V52           V53           V54
#>  [1,]  6.449875e-05 4.543693e-05 2.759946e-05  4.671787e-05  3.309097e-05
#>  [2,]  3.345356e-05 2.490751e-05 2.015607e-05  1.998739e-05  1.864939e-05
#>  [3,] -3.995280e-06 5.255752e-07 4.034803e-06 -8.003099e-06 -1.634034e-06
#>  [4,]  3.150537e-05 1.870169e-05 1.198052e-05  2.486187e-05  1.830196e-05
#>  [5,]  1.721991e-05 1.548897e-05 1.447555e-05  7.365158e-06  1.016872e-05
#>  [6,]  6.335999e-05 4.645013e-05 3.639688e-05  3.975868e-05  3.457093e-05
#>  [7,]  2.058130e-05 2.609365e-05 2.940397e-05 -2.384625e-07  1.335060e-05
#>  [8,]  1.656497e-05 9.413383e-06 6.064215e-06  1.269389e-05  9.504148e-06
#>  [9,]  1.058678e-04 6.965659e-05 4.472583e-05  7.880095e-05  5.593494e-05
#> [10,]  2.698544e-05 2.961475e-05 3.254568e-05  5.082863e-06  1.680269e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  3.032017e-05  3.097397e-05  5.334832e-05  3.121296e-05  2.581185e-05
#>  [2,]  1.486578e-05  1.388851e-05  1.780092e-05  1.439056e-05  1.305116e-05
#>  [3,] -3.144559e-06 -3.318452e-06 -1.134610e-05 -2.131816e-06 -2.127767e-06
#>  [4,]  1.684129e-05  1.662272e-05  2.663515e-05  1.503031e-05  1.403613e-05
#>  [5,]  6.633851e-06  5.419824e-06  2.957422e-06  6.896701e-06  5.921757e-06
#>  [6,]  2.819649e-05  2.576920e-05  3.635753e-05  2.775737e-05  2.405391e-05
#>  [7,]  5.769790e-06 -9.194650e-08 -1.695615e-05  5.495597e-06  4.768205e-06
#>  [8,]  8.872343e-06  9.359896e-06  1.470777e-05  8.330225e-06  7.697351e-06
#>  [9,]  4.939826e-05  5.009059e-05  8.738919e-05  5.106829e-05  4.252648e-05
#> [10,]  9.240555e-06  2.469581e-06 -1.178379e-05  8.391821e-06  6.982176e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  3.414865e-04  2.650368e-05  4.838209e-04  5.585445e-04  7.089496e-04
#>  [2,]  1.502298e-04  1.127675e-05  2.053826e-04  2.346082e-04  3.248217e-04
#>  [3,] -4.193671e-05 -5.683427e-06 -6.933716e-05 -4.480711e-05 -4.728431e-06
#>  [4,]  1.688125e-04  1.625391e-05  2.463425e-04  2.242813e-04  2.443700e-04
#>  [5,]  6.252593e-05  2.907429e-06  7.561284e-05  1.155997e-04  1.968130e-04
#>  [6,]  2.906908e-04  2.136706e-05  3.915610e-04  4.573468e-04  6.242479e-04
#>  [7,]  1.370411e-05 -4.593668e-06 -3.247203e-05  6.238602e-05  2.230223e-04
#>  [8,]  9.301197e-05  8.704847e-06  1.342918e-04  1.162185e-04  1.216351e-04
#>  [9,]  5.641011e-04  4.499227e-05  7.921499e-04  8.621180e-04  1.041376e-03
#> [10,]  3.436287e-05 -1.776684e-06  1.300736e-05  1.259996e-04  3.103754e-04
#> 
#> $vW[[2]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]       [,6]       [,7]
#> [1,]  0.2286892  0.158834  0.1489231  0.102999  0.1180903  0.1754888  0.2006569
#> [2,] -0.2286892 -0.158834 -0.1489231 -0.102999 -0.1180903 -0.1754888 -0.2006569
#>            [,8]       [,9]      [,10]
#> [1,]  0.1078893  0.2002575  0.1963368
#> [2,] -0.1078893 -0.2002575 -0.1963368
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.8575847 0.3583472 0.3906700 0.2491401 0.2777624 0.5607211 0.7720571
#>  [8] 0.1678907 0.6338002 0.7217490
#> 
#> $B[[2]]
#> [1] -0.19061620  0.07735033
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0060104196  0.0020789014 -0.0013205167  0.0031425020  0.0004372979
#>  [6]  0.0042085062 -0.0017022621  0.0016449619  0.0098193162 -0.0010835268
#> 
#> $vB[[2]]
#> [1]  0.2492496 -0.2492496
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#>  [2,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#>  [3,] 0.0025 0.1430 0.0994 0.2250 0.2444 0.3239 0.3039 0.2410 0.0367 0.1672
#>  [4,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#>  [5,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#>  [6,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#>  [7,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#>  [8,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>  [9,] 0.0208 0.1093 0.1063 0.1179 0.1291 0.1591 0.1680 0.1918 0.1615 0.1647
#> [10,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [11,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [12,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [13,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [14,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [15,] 0.0378 0.2053 0.3135 0.3118 0.3686 0.3885 0.5850 0.7868 0.9739 1.0000
#> [16,] 0.0117 0.2169 0.1180 0.0754 0.2782 0.3758 0.5093 0.6592 0.7071 0.7532
#> [17,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [18,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [19,] 0.0123 0.0835 0.0548 0.0847 0.2026 0.2557 0.1870 0.2032 0.1463 0.2849
#> [20,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [21,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [22,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [23,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [24,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [25,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [26,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [27,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [28,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [29,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#> [30,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [31,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [32,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [33,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [34,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [35,] 0.0519 0.2838 0.2802 0.3086 0.2657 0.3801 0.5626 0.4376 0.2617 0.1199
#> [36,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [37,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [38,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [39,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [40,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#>  [2,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#>  [3,] 0.3038 0.0309 0.4069 0.3613 0.1994 0.4611 0.6849 0.7272 0.7152 0.7102
#>  [4,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#>  [5,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#>  [6,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#>  [7,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#>  [8,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>  [9,] 0.1397 0.0186 0.1426 0.2429 0.2816 0.4290 0.6443 0.9061 1.0000 0.8087
#> [10,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [11,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [12,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [13,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [14,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [15,] 0.9843 0.0318 0.8610 0.8443 0.9061 0.5847 0.4033 0.5946 0.6793 0.6389
#> [16,] 0.8357 0.0069 0.8593 0.9615 0.9838 0.8705 0.6403 0.5067 0.5395 0.6934
#> [17,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [18,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [19,] 0.5824 0.0309 0.7728 0.7852 0.8515 0.5312 0.3653 0.5973 0.8275 1.0000
#> [20,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [21,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [22,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [23,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [24,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [25,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [26,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [27,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [28,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [29,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#> [30,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [31,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [32,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [33,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [34,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [35,] 0.6676 0.0548 0.9402 0.7832 0.5352 0.6809 0.9174 0.7613 0.8220 0.8872
#> [36,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [37,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [38,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [39,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [40,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#>  [2,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#>  [3,] 0.8516 1.0000 0.0171 0.7690 0.4841 0.3717 0.6096 0.5110 0.2586 0.0916
#>  [4,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#>  [5,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#>  [6,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#>  [7,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#>  [8,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>  [9,] 0.6119 0.5260 0.0131 0.3677 0.2746 0.1020 0.1339 0.1582 0.1952 0.1787
#> [10,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [11,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [12,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [13,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [14,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [15,] 0.5002 0.5578 0.0423 0.4831 0.4729 0.3318 0.3969 0.3894 0.2314 0.1036
#> [16,] 0.8487 0.8213 0.0279 0.5962 0.2950 0.2758 0.2885 0.1893 0.1446 0.0955
#> [17,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [18,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [19,] 0.8673 0.6301 0.0169 0.4591 0.3940 0.2576 0.2817 0.2641 0.2757 0.2698
#> [20,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [21,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [22,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [23,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [24,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [25,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [26,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [27,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [28,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [29,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#> [30,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [31,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [32,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [33,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [34,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [35,] 0.6091 0.2967 0.0842 0.1103 0.1318 0.0624 0.0990 0.4006 0.3666 0.1050
#> [36,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [37,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [38,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [39,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [40,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#>  [2,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#>  [3,] 0.0947 0.2287 0.3480 0.0228 0.2095 0.1901 0.2941 0.2211 0.1524 0.0746
#>  [4,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#>  [5,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#>  [6,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#>  [7,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#>  [8,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>  [9,] 0.0429 0.1096 0.1762 0.0211 0.2481 0.3150 0.2920 0.1902 0.0696 0.0758
#> [10,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [11,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [12,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [13,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [14,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [15,] 0.1312 0.0864 0.2569 0.0350 0.3179 0.2649 0.2714 0.1713 0.0584 0.1230
#> [16,] 0.0888 0.0836 0.0894 0.0583 0.1547 0.2318 0.2225 0.1035 0.1721 0.2017
#> [17,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [18,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [19,] 0.3994 0.4576 0.3940 0.0313 0.2522 0.1782 0.1354 0.0516 0.0337 0.0894
#> [20,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [21,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [22,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [23,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [24,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [25,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [26,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [27,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [28,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [29,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#> [30,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [31,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [32,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [33,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [34,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [35,] 0.1915 0.3930 0.4288 0.0319 0.2546 0.1151 0.2196 0.1879 0.1437 0.2146
#> [36,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [37,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [38,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [39,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [40,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#>  [2,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#>  [3,] 0.0606 0.0692 0.0446 0.0344 0.0434 0.0082 0.0108 0.0149 0.0077 0.0036
#>  [4,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#>  [5,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#>  [6,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#>  [7,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#>  [8,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>  [9,] 0.0910 0.0441 0.0244 0.0265 0.0610 0.0095 0.0140 0.0074 0.0063 0.0081
#> [10,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [11,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [12,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [13,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [14,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [15,] 0.2200 0.2198 0.1074 0.0423 0.1787 0.0162 0.0093 0.0046 0.0044 0.0078
#> [16,] 0.1787 0.1112 0.0398 0.0305 0.0915 0.0084 0.0039 0.0053 0.0029 0.0020
#> [17,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [18,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [19,] 0.0861 0.0872 0.0445 0.0134 0.0358 0.0217 0.0188 0.0133 0.0265 0.0224
#> [20,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [21,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [22,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [23,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [24,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [25,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [26,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [27,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [28,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [29,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#> [30,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [31,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [32,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [33,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [34,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [35,] 0.2360 0.1125 0.0254 0.0285 0.1158 0.0178 0.0052 0.0081 0.0120 0.0045
#> [36,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [37,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [38,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [39,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [40,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#>  [2,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#>  [3,] 0.0114 0.0085 0.0101 0.0016 0.0028 0.1224 0.0014 0.1947 0.1661 0.1368
#>  [4,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#>  [5,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#>  [6,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#>  [7,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#>  [8,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#>  [9,] 0.0087 0.0044 0.0028 0.0019 0.0049 0.0613 0.0023 0.0612 0.0506 0.0989
#> [10,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [11,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [12,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [13,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [14,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [15,] 0.0102 0.0065 0.0061 0.0062 0.0043 0.1635 0.0053 0.0887 0.0817 0.1779
#> [16,] 0.0013 0.0029 0.0020 0.0062 0.0026 0.1267 0.0052 0.1577 0.1927 0.2361
#> [17,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [18,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [19,] 0.0074 0.0118 0.0026 0.0092 0.0009 0.0102 0.0044 0.0182 0.0579 0.1122
#> [20,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [21,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [22,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [23,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [24,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [25,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [26,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [27,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [28,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [29,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#> [30,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [31,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [32,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [33,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [34,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [35,] 0.0121 0.0097 0.0085 0.0047 0.0048 0.0922 0.0053 0.1027 0.0613 0.1465
#> [36,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [37,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [38,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [39,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [40,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9259884 0.6243411 0.5357694 0.3709813 0.3970652 0.6641167 0.8105458
#>  [2,] 0.8389226 0.4722907 0.3743275 0.2497883 0.3377834 0.5323027 0.6792944
#>  [3,] 0.9163664 0.6360760 0.5795188 0.3970809 0.4426426 0.6998744 0.7860726
#>  [4,] 0.8776427 0.5141020 0.5155268 0.3440651 0.4571387 0.6597662 0.7180227
#>  [5,] 0.8860372 0.4903486 0.4695650 0.3213974 0.3607221 0.6305464 0.6785132
#>  [6,] 0.8904024 0.4885212 0.4665095 0.3250520 0.4819624 0.6803063 0.6900319
#>  [7,] 0.9221710 0.6763311 0.6265999 0.4550460 0.4518084 0.7256872 0.8171464
#>  [8,] 0.8932563 0.5286707 0.4276877 0.2455060 0.3473106 0.5707150 0.7542852
#>  [9,] 0.8918978 0.6505454 0.6438441 0.4898343 0.5068055 0.7394629 0.7686662
#> [10,] 0.8803507 0.5717784 0.4122136 0.2725585 0.3420640 0.5293868 0.7677698
#> [11,] 0.9442667 0.6723126 0.6844903 0.5219773 0.5328006 0.7944066 0.8119745
#> [12,] 0.8688796 0.3985638 0.3884477 0.1997817 0.3894843 0.5689896 0.6910545
#> [13,] 0.9004246 0.5255155 0.5256547 0.3137782 0.4239314 0.7024186 0.7467380
#> [14,] 0.8966055 0.4633178 0.5083909 0.2997726 0.5004826 0.7065211 0.7273620
#> [15,] 0.9021200 0.5252614 0.5202513 0.3359047 0.4863652 0.6899741 0.7589601
#> [16,] 0.9260043 0.5828710 0.5850378 0.4061090 0.5279476 0.7467289 0.7824377
#> [17,] 0.9366038 0.6509913 0.6103030 0.3998354 0.4690090 0.7513671 0.7977077
#> [18,] 0.8902907 0.6544982 0.6322196 0.4914293 0.5027137 0.7306435 0.7728442
#> [19,] 0.9239577 0.6476835 0.6366711 0.4666134 0.5509812 0.7461956 0.7903718
#> [20,] 0.9452407 0.6182373 0.6555782 0.4524716 0.5754648 0.8001105 0.8189648
#> [21,] 0.8900387 0.5870062 0.5800526 0.4088525 0.5013158 0.7103917 0.7395084
#> [22,] 0.9254508 0.6145806 0.5176828 0.3413410 0.3716308 0.6386829 0.8067101
#> [23,] 0.8840329 0.5175284 0.5228398 0.3490931 0.4627860 0.6988430 0.7355930
#> [24,] 0.9134686 0.6442103 0.6224002 0.4657414 0.5088819 0.7428072 0.7745849
#> [25,] 0.9589309 0.6782190 0.6370343 0.4116700 0.4866478 0.7471323 0.8436625
#> [26,] 0.8647801 0.4572267 0.3154365 0.1972407 0.3142920 0.4673308 0.7110604
#> [27,] 0.9193786 0.5707408 0.5832238 0.3754166 0.5274691 0.7358637 0.7928931
#> [28,] 0.8956573 0.4895433 0.4773143 0.2844562 0.3893918 0.6697143 0.7257422
#> [29,] 0.9466919 0.6696321 0.6212760 0.4138063 0.4452101 0.7315504 0.8329924
#> [30,] 0.8414553 0.4292247 0.4202308 0.2633223 0.5016892 0.6014173 0.6867015
#> [31,] 0.9226769 0.5405492 0.5283510 0.3346589 0.4777201 0.7010442 0.7508594
#> [32,] 0.9462284 0.6298124 0.6373416 0.4047519 0.5167967 0.7582760 0.8069990
#> [33,] 0.8888393 0.4761335 0.4729447 0.3236200 0.4614754 0.6259036 0.7143637
#> [34,] 0.9295039 0.5418146 0.5687242 0.3399562 0.4949011 0.7427936 0.7640340
#> [35,] 0.8947363 0.5676619 0.5662848 0.3747028 0.5400150 0.7191652 0.7403401
#> [36,] 0.8904760 0.5851711 0.5386332 0.3916382 0.5017444 0.6693282 0.7431668
#> [37,] 0.8839642 0.5543134 0.4269064 0.2697774 0.3548530 0.5408530 0.7691789
#> [38,] 0.9333136 0.6305907 0.6475310 0.4680615 0.4940658 0.7542379 0.7946815
#> [39,] 0.9340066 0.5879949 0.6170827 0.4118531 0.5121562 0.7728583 0.7827879
#> [40,] 0.9554470 0.6791833 0.6648259 0.4360266 0.5014656 0.7647402 0.8399511
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4023156 0.7898044 0.7796918
#>  [2,] 0.2544517 0.6426632 0.6931500
#>  [3,] 0.4080238 0.7967751 0.7831070
#>  [4,] 0.3370177 0.7361626 0.7440737
#>  [5,] 0.2887431 0.7284516 0.7072816
#>  [6,] 0.3038275 0.7723032 0.7351030
#>  [7,] 0.4707340 0.7981929 0.8057008
#>  [8,] 0.2736489 0.7062540 0.7446906
#>  [9,] 0.4784921 0.7912139 0.7867111
#> [10,] 0.3184438 0.6732255 0.7394166
#> [11,] 0.5123629 0.8557356 0.8200087
#> [12,] 0.2280989 0.7179469 0.6848306
#> [13,] 0.3055037 0.7755941 0.7651188
#> [14,] 0.2875295 0.7900471 0.7315243
#> [15,] 0.3375312 0.7942240 0.7572826
#> [16,] 0.3951894 0.8337716 0.7890924
#> [17,] 0.3951387 0.8426727 0.8075631
#> [18,] 0.4838236 0.7855870 0.7882714
#> [19,] 0.4728526 0.8309624 0.7921104
#> [20,] 0.4534285 0.8724287 0.8218126
#> [21,] 0.3923891 0.7933438 0.7701536
#> [22,] 0.3763440 0.7820609 0.7763431
#> [23,] 0.3284961 0.7790762 0.7495554
#> [24,] 0.4548669 0.8099009 0.7960901
#> [25,] 0.4472794 0.8570379 0.8288316
#> [26,] 0.2252398 0.6569371 0.6933588
#> [27,] 0.3791072 0.8250707 0.7784336
#> [28,] 0.2849409 0.7595744 0.7437585
#> [29,] 0.4381323 0.8245022 0.8212993
#> [30,] 0.2638980 0.7260738 0.6999157
#> [31,] 0.3324037 0.7897293 0.7890053
#> [32,] 0.4177542 0.8422909 0.8120513
#> [33,] 0.3232813 0.7352473 0.7299872
#> [34,] 0.3301335 0.8312602 0.7817408
#> [35,] 0.3799288 0.7878469 0.7654454
#> [36,] 0.3918905 0.7751254 0.7630059
#> [37,] 0.3194129 0.7106557 0.7099660
#> [38,] 0.4726195 0.8345123 0.7990420
#> [39,] 0.3949003 0.8437416 0.8013263
#> [40,] 0.4631713 0.8595258 0.8278803
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6533221 0.3466779
#>  [2,] 0.6399408 0.3600592
#>  [3,] 0.6539205 0.3460795
#>  [4,] 0.6482972 0.3517028
#>  [5,] 0.6489242 0.3510758
#>  [6,] 0.6516761 0.3483239
#>  [7,] 0.6547101 0.3452899
#>  [8,] 0.6453836 0.3546164
#>  [9,] 0.6537157 0.3462843
#> [10,] 0.6429381 0.3570619
#> [11,] 0.6602857 0.3397143
#> [12,] 0.6450142 0.3549858
#> [13,] 0.6512142 0.3487858
#> [14,] 0.6519949 0.3480051
#> [15,] 0.6523610 0.3476390
#> [16,] 0.6567268 0.3432732
#> [17,] 0.6575888 0.3424112
#> [18,] 0.6532685 0.3467315
#> [19,] 0.6571962 0.3428038
#> [20,] 0.6600589 0.3399411
#> [21,] 0.6528664 0.3471336
#> [22,] 0.6523062 0.3476938
#> [23,] 0.6510786 0.3489214
#> [24,] 0.6556501 0.3443499
#> [25,] 0.6589804 0.3410196
#> [26,] 0.6405882 0.3594118
#> [27,] 0.6552255 0.3447745
#> [28,] 0.6500145 0.3499855
#> [29,] 0.6566833 0.3433167
#> [30,] 0.6446928 0.3553072
#> [31,] 0.6533550 0.3466450
#> [32,] 0.6578927 0.3421073
#> [33,] 0.6485553 0.3514447
#> [34,] 0.6560960 0.3439040
#> [35,] 0.6526187 0.3473813
#> [36,] 0.6515266 0.3484734
#> [37,] 0.6453449 0.3546551
#> [38,] 0.6578221 0.3421779
#> [39,] 0.6578184 0.3421816
#> [40,] 0.6593164 0.3406836
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]        [,2]        [,3]        [,4]         [,5]       [,6]
#>  [1,] 2.526640  0.50801517  0.14332231 -0.52800945 -0.417708453  0.6816941
#>  [2,] 1.650233 -0.11095073 -0.51369609 -1.09974156 -0.673187922  0.1293910
#>  [3,] 2.393970  0.55837295  0.32079825 -0.41764300 -0.230444095  0.8466997
#>  [4,] 1.970294  0.05642300  0.06212717 -0.64523064 -0.171867016  0.6622524
#>  [5,] 2.050887 -0.03861034 -0.12189055 -0.74735725 -0.572231484  0.5345615
#>  [6,] 2.094858 -0.04592317 -0.13416305 -0.73065029 -0.072181817  0.7551800
#>  [7,] 2.472217  0.73696162  0.51765787 -0.18030295 -0.193366711  0.9728498
#>  [8,] 2.124443  0.11480889 -0.29129174 -1.12272618 -0.630882100  0.2847689
#>  [9,] 2.110275  0.62143758  0.59208822 -0.04066858  0.027223641  1.0431788
#> [10,] 1.995755  0.28911075 -0.35482176 -0.98168030 -0.654109926  0.1176830
#> [11,] 2.829830  0.71866326  0.77448515  0.08796593  0.131390924  1.3516951
#> [12,] 1.891089 -0.41145277 -0.45384211 -1.38765923 -0.449480623  0.2777301
#> [13,] 2.201952  0.10215063  0.10270900 -0.78251437 -0.306655080  0.8588418
#> [14,] 2.160064 -0.14699299  0.03356656 -0.84838102  0.001930544  0.8785471
#> [15,] 2.221005  0.10113183  0.08104970 -0.68159821 -0.054552739  0.7999983
#> [16,] 2.526872  0.33457041  0.34348885 -0.38007424  0.111907154  1.0812416
#> [17,] 2.692856  0.62339955  0.44858585 -0.40615084 -0.124122993  1.1059166
#> [18,] 2.093714  0.63887080  0.54175066 -0.03428611  0.010854712  0.9978898
#> [19,] 2.497377  0.60887243  0.56094459 -0.13374535  0.204635820  1.0784237
#> [20,] 2.848492  0.48207298  0.64365071 -0.19068950  0.304183296  1.3869850
#> [21,] 2.091136  0.35160282  0.32298919 -0.36871092  0.005263265  0.8972874
#> [22,] 2.518821  0.46660786  0.07076067 -0.65732397 -0.525226881  0.5696521
#> [23,] 2.031188  0.07014239  0.09142292 -0.62302781 -0.149131700  0.8417946
#> [24,] 2.356741  0.59368550  0.49974844 -0.13724960  0.035531162  1.0606104
#> [25,] 3.150564  0.74559924  0.56251509 -0.35706577 -0.053421506  1.0833760
#> [26,] 1.855573 -0.17151235 -0.77482393 -1.40362980 -0.780129230 -0.1308631
#> [27,] 2.433935  0.28487438  0.33602177 -0.50904856  0.109987013  1.0245795
#> [28,] 2.149877 -0.04183290 -0.09080497 -0.92246366 -0.449869440  0.7068930
#> [29,] 2.876885  0.70652180  0.49496792 -0.34825210 -0.220043177  1.0025027
#> [30,] 1.669096 -0.28501505 -0.32182594 -1.02877175  0.006756821  0.4113741
#> [31,] 2.479287  0.16255381  0.11352575 -0.68718797 -0.089178785  0.8522751
#> [32,] 2.867739  0.53141196  0.56384427 -0.38570403  0.067211913  1.1432511
#> [33,] 2.078940 -0.09553857 -0.10832704 -0.73718499 -0.154404320  0.5146827
#> [34,] 2.579093  0.16764989  0.27664763 -0.66348934 -0.020396121  1.0605392
#> [35,] 2.140060  0.27231796  0.26670912 -0.51209394  0.160402975  0.9403245
#> [36,] 2.095613  0.34403809  0.15484148 -0.44043114  0.006977770  0.7051480
#> [37,] 2.030517  0.21811418 -0.29448441 -0.99575200 -0.597774433  0.1637772
#> [38,] 2.638740  0.53475191  0.60820387 -0.12792834 -0.023737856  1.1213437
#> [39,] 2.649929  0.35568241  0.47718411 -0.35630997  0.048634216  1.2245215
#> [40,] 3.065499  0.75002098  0.68487499 -0.25730393  0.005862461  1.1788457
#>            [,7]        [,8]      [,9]     [,10]
#>  [1,] 1.4535606 -0.39582581 1.3237466 1.2638712
#>  [2,] 0.7505312 -1.07500883 0.5869422 0.8148876
#>  [3,] 1.3014122 -0.37214072 1.3662594 1.2838649
#>  [4,] 0.9346744 -0.67661264 1.0261182 1.0672503
#>  [5,] 0.7469477 -0.90149635 0.9867807 0.8822179
#>  [6,] 0.8002685 -0.82913742 1.2213623 1.0206695
#>  [7,] 1.4971324 -0.11719778 1.3750380 1.4223130
#>  [8,] 1.1215992 -0.97618784 0.8772595 1.0704927
#>  [9,] 1.2007954 -0.08608479 1.3322581 1.3052137
#> [10,] 1.1957607 -0.76093286 0.7228101 1.0429384
#> [11,] 1.4628916  0.04946170 1.7803136 1.5164066
#> [12,] 0.8050540 -1.21907717 0.9343003 0.7760615
#> [13,] 1.0812897 -0.82122481 1.2401729 1.1809510
#> [14,] 0.9812795 -0.90741298 1.3252092 1.0023698
#> [15,] 1.1469868 -0.67431557 1.3505773 1.1378390
#> [16,] 1.2799289 -0.42555028 1.6125970 1.3194633
#> [17,] 1.3720286 -0.42576245 1.6782506 1.4342530
#> [18,] 1.2244415 -0.06472811 1.2985272 1.3145371
#> [19,] 1.3271681 -0.10869642 1.5924635 1.3376939
#> [20,] 1.5093495 -0.18682743 1.9226056 1.5286764
#> [21,] 1.0434149 -0.43728091 1.3451998 1.2091785
#> [22,] 1.4287730 -0.50509550 1.2777169 1.2444813
#> [23,] 1.0231875 -0.71499485 1.2602909 1.0962427
#> [24,] 1.2343839 -0.18102521 1.4493662 1.3620340
#> [25,] 1.6857352 -0.21166910 1.7909028 1.5773690
#> [26,] 0.9005399 -1.23538785 0.6496744 0.8158691
#> [27,] 1.3424535 -0.49333938 1.5510871 1.2565615
#> [28,] 0.9731264 -0.92008316 1.1503478 1.0655958
#> [29,] 1.6069849 -0.24874542 1.5471530 1.5251756
#> [30,] 0.7847431 -1.02580583 0.9747930 0.8468964
#> [31,] 1.1032008 -0.69733334 1.3232947 1.3189402
#> [32,] 1.4306271 -0.33199937 1.6753730 1.4633944
#> [33,] 0.9166728 -0.73873293 1.0214108 0.9945579
#> [34,] 1.1749246 -0.70758138 1.5945851 1.2758397
#> [35,] 1.0477370 -0.48985028 1.3119956 1.1827697
#> [36,] 1.0624936 -0.43937267 1.2374816 1.1692302
#> [37,] 1.2036804 -0.75647139 0.8985708 0.8952191
#> [38,] 1.3533793 -0.10963163 1.6179508 1.3803173
#> [39,] 1.2819877 -0.42676009 1.6863355 1.3946045
#> [40,] 1.6578644 -0.14758228 1.8113569 1.5706782
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.2841094 -0.3495650
#>  [2,] 0.2487087 -0.3263984
#>  [3,] 0.2902169 -0.3461005
#>  [4,] 0.2767607 -0.3348022
#>  [5,] 0.2691006 -0.3452132
#>  [6,] 0.2870412 -0.3393738
#>  [7,] 0.2943083 -0.3455001
#>  [8,] 0.2591209 -0.3396875
#>  [9,] 0.2980654 -0.3373473
#> [10,] 0.2567218 -0.3314174
#> [11,] 0.3138916 -0.3506758
#> [12,] 0.2587896 -0.3384048
#> [13,] 0.2781273 -0.3462532
#> [14,] 0.2852958 -0.3425240
#> [15,] 0.2870090 -0.3424246
#> [16,] 0.3021228 -0.3466189
#> [17,] 0.2995746 -0.3529930
#> [18,] 0.2972896 -0.3361478
#> [19,] 0.3085021 -0.3423220
#> [20,] 0.3137640 -0.3497929
#> [21,] 0.2927586 -0.3389041
#> [22,] 0.2784817 -0.3507100
#> [23,] 0.2825881 -0.3411956
#> [24,] 0.3017086 -0.3422605
#> [25,] 0.3040562 -0.3546975
#> [26,] 0.2456505 -0.3322676
#> [27,] 0.2974025 -0.3446865
#> [28,] 0.2721648 -0.3469381
#> [29,] 0.2956918 -0.3528568
#> [30,] 0.2710348 -0.3247565
#> [31,] 0.2887042 -0.3451151
#> [32,] 0.3032713 -0.3506460
#> [33,] 0.2767251 -0.3359698
#> [34,] 0.2946286 -0.3513159
#> [35,] 0.2954455 -0.3351243
#> [36,] 0.2907399 -0.3350164
#> [37,] 0.2616119 -0.3370273
#> [38,] 0.3030676 -0.3505363
#> [39,] 0.3021829 -0.3514045
#> [40,] 0.3062098 -0.3540396
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.6533221  0.6533221
#>  [2,]  0.3600592 -0.3600592
#>  [3,] -0.6539205  0.6539205
#>  [4,] -0.6482972  0.6482972
#>  [5,]  0.3510758 -0.3510758
#>  [6,]  0.3483239 -0.3483239
#>  [7,] -0.6547101  0.6547101
#>  [8,] -0.6453836  0.6453836
#>  [9,] -0.6537157  0.6537157
#> [10,] -0.6429381  0.6429381
#> [11,] -0.6602857  0.6602857
#> [12,] -0.6450142  0.6450142
#> [13,] -0.6512142  0.6512142
#> [14,]  0.3480051 -0.3480051
#> [15,] -0.6523610  0.6523610
#> [16,]  0.3432732 -0.3432732
#> [17,] -0.6575888  0.6575888
#> [18,] -0.6532685  0.6532685
#> [19,] -0.6571962  0.6571962
#> [20,] -0.6600589  0.6600589
#> [21,]  0.3471336 -0.3471336
#> [22,] -0.6523062  0.6523062
#> [23,] -0.6510786  0.6510786
#> [24,] -0.6556501  0.6556501
#> [25,] -0.6589804  0.6589804
#> [26,] -0.6405882  0.6405882
#> [27,] -0.6552255  0.6552255
#> [28,] -0.6500145  0.6500145
#> [29,] -0.6566833  0.6566833
#> [30,]  0.3553072 -0.3553072
#> [31,]  0.3466450 -0.3466450
#> [32,]  0.3421073 -0.3421073
#> [33,] -0.6485553  0.6485553
#> [34,]  0.3439040 -0.3439040
#> [35,] -0.6526187  0.6526187
#> [36,] -0.6515266  0.6515266
#> [37,] -0.6453449  0.6453449
#> [38,]  0.3421779 -0.3421779
#> [39,]  0.3421816 -0.3421816
#> [40,] -0.6593164  0.6593164
#> 
#> $L
#> [1] 0.6899892 0.7026851 0.7421846 0.7166144 0.6822372 0.8678660
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5362319