Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10         V11          V12         V13         V14
#>  [1,] -0.2348910 -0.008809212 -0.07387139 -0.053437396 -0.09589560  0.02700271
#>  [2,] -0.2133292 -0.164711606 -0.16431096 -0.135099335 -0.02625915  0.03349494
#>  [3,] -0.3420350 -0.097805589 -0.13942307 -0.074476163 -0.13983962 -0.14929487
#>  [4,] -0.3084635 -0.086435581 -0.20558092 -0.174099117 -0.06934074  0.05609794
#>  [5,] -0.2120435 -0.131998306 -0.08527763 -0.010371016 -0.04565501 -0.02172417
#>  [6,] -0.2641005 -0.196820065 -0.12362239 -0.073115028 -0.13384268 -0.17735951
#>  [7,] -0.3280960 -0.236924916 -0.14455422 -0.053427163 -0.20389707 -0.02807724
#>  [8,] -0.3168266 -0.091653104 -0.02008300 -0.099481865  0.02500168 -0.13250430
#>  [9,] -0.4100418 -0.205528659 -0.08369176 -0.103856883 -0.06182973 -0.01054892
#> [10,] -0.3686225 -0.104993253 -0.13229686 -0.005332594 -0.11820354 -0.08130274
#>               V15         V16         V17          V18         V19         V2
#>  [1,] -0.05976989  0.02940351  0.11328488  0.040786390  0.06964361 -0.3234833
#>  [2,]  0.04315120 -0.02509920  0.04586660  0.096950899  0.15165068 -0.2457421
#>  [3,] -0.04110911 -0.08184251 -0.07265869  0.089714480 -0.01379241 -0.3223678
#>  [4,] -0.01374209  0.06607356 -0.04071780 -0.006312404  0.01676323 -0.2370061
#>  [5,] -0.04888202  0.01220921  0.08114090  0.016020772  0.09987614 -0.3631498
#>  [6,] -0.13113885 -0.22631172 -0.06918848 -0.045078584 -0.04353134 -0.3465784
#>  [7,]  0.01821611 -0.13179914 -0.03421623 -0.024583133 -0.04281225 -0.2823924
#>  [8,] -0.13394648  0.04053970 -0.03368210  0.106153285  0.12441500 -0.3167672
#>  [9,] -0.04789837  0.01800166  0.08914328  0.111983323  0.20958912 -0.3152788
#> [10,]  0.06296422  0.08287251  0.06113820  0.076422437  0.14711521 -0.2829000
#>                V20         V21         V22        V23        V24        V25
#>  [1,]  0.190941353  0.07816641  0.07859321 0.18385830 0.10848276 0.11626448
#>  [2,]  0.133846885  0.15681167  0.21108869 0.27845259 0.12531690 0.11944574
#>  [3,]  0.022997751  0.05135546  0.02401457 0.12308683 0.08462376 0.12224525
#>  [4,] -0.001854780  0.10394054  0.09772398 0.04141678 0.22180370 0.14021760
#>  [5,]  0.211570022  0.19533494  0.14570022 0.27941293 0.17642502 0.24561750
#>  [6,] -0.008743947  0.01224196 -0.04363618 0.05595345 0.12513648 0.04632691
#>  [7,] -0.034949972 -0.05102703  0.04232973 0.07244977 0.01993847 0.05757593
#>  [8,]  0.211857780  0.20290984  0.20961167 0.12601383 0.11536477 0.24733033
#>  [9,]  0.233483633  0.20153127  0.27722441 0.22138545 0.24559311 0.17017918
#> [10,]  0.165743629  0.25176673  0.22049594 0.14654808 0.34277983 0.33421031
#>              V26        V27         V28         V29         V3          V30
#>  [1,] 0.03425454 0.06193369  0.09328072  0.12511876 -0.2251732  0.064224952
#>  [2,] 0.26387634 0.16897593  0.09520806  0.07143142 -0.3521241 -0.005843377
#>  [3,] 0.16608221 0.18266542  0.14029893 -0.02166139 -0.3021912 -0.028101707
#>  [4,] 0.15275512 0.09313502  0.07224582 -0.02522213 -0.3309339  0.013958392
#>  [5,] 0.18640360 0.14525091  0.15042832 -0.03111502 -0.2389064 -0.039874129
#>  [6,] 0.11698278 0.02082535  0.16433274  0.14300503 -0.2580015 -0.021038773
#>  [7,] 0.05346556 0.08632999 -0.02624418 -0.04037930 -0.2191385 -0.033525013
#>  [8,] 0.19691494 0.18515827  0.27558911  0.16977297 -0.3159151  0.075331763
#>  [9,] 0.20953478 0.11649416  0.09911018  0.11894137 -0.2765752  0.007194375
#> [10,] 0.23083378 0.34612867  0.21298083  0.21271904 -0.3189349  0.256280960
#>               V31         V32          V33          V34         V35
#>  [1,]  0.14265692 -0.01428495 -0.015712242  0.055736393 -0.01302916
#>  [2,] -0.04642755 -0.03286557 -0.071646285 -0.150672705 -0.07848980
#>  [3,]  0.02688958 -0.05925619 -0.101563001 -0.075298634 -0.06579555
#>  [4,]  0.01552864 -0.08548524  0.011976960 -0.173448815 -0.18136355
#>  [5,] -0.04388592 -0.06532553 -0.049149767 -0.005003447 -0.13500259
#>  [6,]  0.04812844  0.03800699  0.001676156 -0.032997373 -0.13902349
#>  [7,] -0.02940160 -0.18331448 -0.172748618 -0.057646116 -0.21254571
#>  [8,]  0.18275324  0.13934543  0.069824450 -0.073706865 -0.05340828
#>  [9,] -0.02810852  0.02193774 -0.074416848  0.082826855  0.08713401
#> [10,]  0.05909026  0.18785620  0.099961705  0.051156519  0.09893414
#>                V36           V37          V38         V39         V4
#>  [1,]  0.031065729 -0.0449018247 -0.022734683 -0.10437048 -0.2664598
#>  [2,] -0.168252039 -0.1168384872 -0.011441039 -0.05002079 -0.3161108
#>  [3,] -0.043317080  0.0344144969 -0.001394916  0.06573987 -0.2464287
#>  [4,] -0.096030216 -0.1388821594 -0.073842386 -0.06184061 -0.3133718
#>  [5,] -0.102838660 -0.1341445769 -0.126039495 -0.03376232 -0.1683420
#>  [6,] -0.001648644 -0.0751336033 -0.094758285 -0.02279940 -0.2617197
#>  [7,] -0.093485856 -0.0537967514 -0.042140853 -0.10301568 -0.3462307
#>  [8,] -0.049839166  0.0017669898  0.019339232 -0.00319277 -0.2435152
#>  [9,]  0.008657988 -0.0001528357 -0.042917181 -0.01721372 -0.2364653
#> [10,]  0.095754128  0.0533884294 -0.080599881 -0.10691061 -0.2968862
#>               V40         V41         V42          V43         V44         V45
#>  [1,] -0.02579340  0.03201271  0.02376144 -0.070893651 -0.09884296 -0.07339109
#>  [2,] -0.08860585 -0.06714878 -0.08340529 -0.002339577 -0.18233975 -0.18988335
#>  [3,]  0.03803808 -0.01155320  0.01397576 -0.161681240 -0.12749655 -0.06289971
#>  [4,] -0.06228557 -0.13566030 -0.09019422 -0.150666294 -0.25822352 -0.20737312
#>  [5,] -0.05014595 -0.12773702 -0.16348461 -0.204637420 -0.08763334 -0.07511392
#>  [6,] -0.03715064 -0.20826847 -0.15694221 -0.134107929 -0.12022227 -0.16447837
#>  [7,] -0.05994995 -0.19142719 -0.23497954 -0.145509238 -0.12039030 -0.22496284
#>  [8,] -0.02222184 -0.10399183 -0.02077986 -0.142116780 -0.10865832 -0.14717198
#>  [9,] -0.05010551 -0.04085293 -0.07059615 -0.122204098 -0.09528300 -0.12094635
#> [10,] -0.07129957  0.06171262 -0.03865440  0.019688445 -0.06280328 -0.13588462
#>               V46        V47        V48        V49         V5        V50
#>  [1,] -0.17901039 -0.2189790 -0.2247555 -0.3365219 -0.1672807 -0.3330654
#>  [2,] -0.12537159 -0.2256793 -0.1477746 -0.3053005 -0.1494076 -0.2229351
#>  [3,] -0.13651251 -0.2449680 -0.2326816 -0.2264282 -0.1823126 -0.3330830
#>  [4,] -0.13580109 -0.1552561 -0.1781528 -0.2506857 -0.2870765 -0.3080523
#>  [5,] -0.16067391 -0.2327179 -0.2389582 -0.2251892 -0.2318984 -0.2536309
#>  [6,] -0.26373575 -0.2553016 -0.2442391 -0.2304872 -0.3878321 -0.3620303
#>  [7,] -0.15291291 -0.1836559 -0.3046824 -0.3503666 -0.1739376 -0.2206609
#>  [8,] -0.19045817 -0.1278115 -0.2885644 -0.2648663 -0.2151492 -0.2903270
#>  [9,] -0.12209277 -0.2430595 -0.3155730 -0.3066078 -0.3573105 -0.3012347
#> [10,] -0.07678964 -0.1023860 -0.1306665 -0.2937077 -0.1940252 -0.4225496
#>              V51        V52        V53        V54        V55        V56
#>  [1,] -0.3767292 -0.2781214 -0.3362690 -0.3614323 -0.2943001 -0.3443596
#>  [2,] -0.3380230 -0.3679226 -0.3799619 -0.3261584 -0.3820209 -0.2161182
#>  [3,] -0.3060601 -0.3024510 -0.4226752 -0.2887759 -0.3158685 -0.3169057
#>  [4,] -0.2617427 -0.2297958 -0.2667755 -0.3216210 -0.3582220 -0.2919688
#>  [5,] -0.2333816 -0.2538793 -0.3110446 -0.2327785 -0.2694899 -0.2739533
#>  [6,] -0.2937601 -0.3340416 -0.2635968 -0.2793022 -0.3813939 -0.2469333
#>  [7,] -0.2031367 -0.3174568 -0.2993466 -0.3865399 -0.4035684 -0.2792404
#>  [8,] -0.3836638 -0.3686617 -0.4852356 -0.4757489 -0.3716630 -0.4876748
#>  [9,] -0.3385006 -0.2939516 -0.3335140 -0.3780867 -0.3979114 -0.3273526
#> [10,] -0.4350030 -0.4181117 -0.3394151 -0.4048661 -0.4001217 -0.4150466
#>              V57        V58        V59         V6        V60          V7
#>  [1,] -0.3604934 -0.3604021 -0.3640052 -0.2330546 -0.4081315 -0.22726394
#>  [2,] -0.2232704 -0.2557754 -0.1980211 -0.1964766 -0.3037977 -0.05291642
#>  [3,] -0.3693284 -0.2694217 -0.3085562 -0.1897679 -0.3190786 -0.20886462
#>  [4,] -0.2419803 -0.3641759 -0.2571261 -0.1932831 -0.2826046 -0.25552610
#>  [5,] -0.3227819 -0.2886581 -0.3605116 -0.1091222 -0.2489596 -0.11582317
#>  [6,] -0.3657011 -0.3782435 -0.3690345 -0.1747670 -0.4005492 -0.21212763
#>  [7,] -0.3281564 -0.4195311 -0.4012407 -0.2418581 -0.2635186 -0.21898048
#>  [8,] -0.4193312 -0.4454176 -0.3517046 -0.2599966 -0.4380568 -0.25256455
#>  [9,] -0.3537080 -0.3292218 -0.3565776 -0.2920105 -0.3732487 -0.16959779
#> [10,] -0.4825778 -0.3394815 -0.4010299 -0.2241598 -0.3199695 -0.23623802
#>                V8          V9
#>  [1,] -0.13080059 -0.16489479
#>  [2,] -0.09542143 -0.11228231
#>  [3,] -0.28604395 -0.16266995
#>  [4,] -0.13133292 -0.15687760
#>  [5,] -0.16802816 -0.08619015
#>  [6,] -0.14500296 -0.22524006
#>  [7,] -0.14355088 -0.17041963
#>  [8,] -0.17457043 -0.23759896
#>  [9,] -0.15329060 -0.10526961
#> [10,] -0.12442185 -0.06387938
#> 
#> $W[[2]]
#>              [,1]       [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] -0.088762615 0.08490764 -0.13651137 -0.01533369 -0.02064578 -0.14091113
#> [2,] -0.003382458 0.04617018  0.03069306 -0.02421890  0.07177395  0.09579335
#>             [,7]        [,8]         [,9]      [,10]
#> [1,] -0.07409754  0.05322965  0.026234641 0.04950163
#> [2,]  0.09526366 -0.02769325 -0.001701642 0.03201162
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 1.189392e-04 0.0010159029 0.0010434936 0.0009633059 0.0011512469
#>  [2,] 9.024180e-05 0.0007678029 0.0007716459 0.0007102509 0.0011145958
#>  [3,] 1.332972e-04 0.0011659969 0.0012510044 0.0012061497 0.0011669583
#>  [4,] 5.470983e-05 0.0003597255 0.0003446951 0.0002757348 0.0007012277
#>  [5,] 1.312359e-04 0.0011171920 0.0011591761 0.0010821481 0.0012168463
#>  [6,] 9.474199e-05 0.0009791382 0.0010944417 0.0011150186 0.0007820159
#>  [7,] 4.495270e-05 0.0004842648 0.0005620857 0.0005996011 0.0003849722
#>  [8,] 5.588504e-05 0.0005876223 0.0005417046 0.0004674110 0.0008531265
#>  [9,] 7.259219e-05 0.0006339932 0.0006182581 0.0005502240 0.0009228048
#> [10,] 4.125657e-05 0.0004363511 0.0003993386 0.0003588348 0.0005090040
#>                V14          V15           V16           V17           V18
#>  [1,] 0.0012423575 1.172925e-03  0.0015627074  0.0017963145  0.0019697696
#>  [2,] 0.0014597754 1.591990e-03  0.0023681214  0.0028592640  0.0031508466
#>  [3,] 0.0010104320 7.845791e-04  0.0008478515  0.0008673353  0.0009508518
#>  [4,] 0.0011189074 1.303415e-03  0.0019330120  0.0023615034  0.0026025865
#>  [5,] 0.0012328997 1.108003e-03  0.0014412653  0.0016276861  0.0017673652
#>  [6,] 0.0003619793 6.090900e-05 -0.0002394657 -0.0004485728 -0.0004815672
#>  [7,] 0.0001445944 1.137671e-05 -0.0001523818 -0.0002616971 -0.0002607618
#>  [8,] 0.0011911576 1.314463e-03  0.0020306610  0.0024828116  0.0027418500
#>  [9,] 0.0012441115 1.372337e-03  0.0020322616  0.0024262214  0.0026783999
#> [10,] 0.0006118804 6.082645e-04  0.0009267573  0.0011056480  0.0012520141
#>                 V19           V2          V20          V21          V22
#>  [1,]  2.261606e-03 1.872449e-04 2.511873e-03 0.0026800875 0.0029135876
#>  [2,]  3.265849e-03 1.913666e-04 3.323958e-03 0.0032283115 0.0034889611
#>  [3,]  1.371930e-03 1.679975e-04 1.756285e-03 0.0021063212 0.0022920091
#>  [4,]  2.579590e-03 1.270758e-04 2.542317e-03 0.0024394594 0.0026683538
#>  [5,]  2.110141e-03 1.951093e-04 2.416383e-03 0.0026389887 0.0028592297
#>  [6,] -8.885441e-05 8.882989e-05 2.400685e-04 0.0006663090 0.0007860100
#>  [7,] -6.720808e-05 3.858909e-05 8.339398e-05 0.0002945482 0.0003312703
#>  [8,]  2.789553e-03 1.593392e-04 2.735937e-03 0.0026149514 0.0028938138
#>  [9,]  2.782447e-03 1.606449e-04 2.784441e-03 0.0026836995 0.0029454307
#> [10,]  1.376806e-03 8.566831e-05 1.414471e-03 0.0013948454 0.0015247554
#>                V23           V24           V25           V26           V27
#>  [1,] 3.091415e-03  0.0030912011  0.0031359459  2.993431e-03  0.0030467342
#>  [2,] 4.045330e-03  0.0044147435  0.0045650073  4.079432e-03  0.0040312004
#>  [3,] 2.165489e-03  0.0018967028  0.0018061504  1.922709e-03  0.0020486893
#>  [4,] 3.188152e-03  0.0035561379  0.0037875229  3.450217e-03  0.0033603994
#>  [5,] 2.966526e-03  0.0028989376  0.0029069954  2.827479e-03  0.0029222257
#>  [6,] 3.640971e-04 -0.0001078982 -0.0003392152 -3.776719e-05  0.0001580016
#>  [7,] 7.650146e-05 -0.0001921495 -0.0003420861 -2.185307e-04 -0.0001286424
#>  [8,] 3.438803e-03  0.0038170590  0.0039119666  3.423262e-03  0.0033682530
#>  [9,] 3.445857e-03  0.0037562013  0.0038889113  3.480218e-03  0.0034746748
#> [10,] 1.726607e-03  0.0018243898  0.0018310353  1.670712e-03  0.0017043529
#>                 V28           V29           V3           V30           V31
#>  [1,]  3.079779e-03  2.787349e-03 2.052793e-04  2.577226e-03  0.0023114164
#>  [2,]  3.986283e-03  3.766455e-03 2.345831e-04  3.494232e-03  0.0034147830
#>  [3,]  2.154909e-03  1.808832e-03 1.699918e-04  1.614186e-03  0.0012352135
#>  [4,]  3.195514e-03  3.131393e-03 1.596537e-04  2.943658e-03  0.0028864702
#>  [5,]  3.004446e-03  2.677947e-03 2.106029e-04  2.475594e-03  0.0021700172
#>  [6,]  3.367419e-04  2.948078e-05 7.377403e-05 -9.612905e-05 -0.0004429719
#>  [7,] -2.989396e-05 -1.813587e-04 3.186866e-05 -2.556376e-04 -0.0004137573
#>  [8,]  3.258456e-03  3.111475e-03 1.887158e-04  2.858786e-03  0.0028581186
#>  [9,]  3.450036e-03  3.236322e-03 1.934019e-04  2.999190e-03  0.0029213024
#> [10,]  1.705511e-03  1.606137e-03 9.378826e-05  1.485426e-03  0.0014041132
#>                 V32           V33           V34           V35           V36
#>  [1,]  1.982066e-03  0.0018956361  0.0018696774  0.0018801127  0.0019994705
#>  [2,]  2.671016e-03  0.0027840205  0.0031308364  0.0034469167  0.0040258241
#>  [3,]  1.273159e-03  0.0010247857  0.0007178643  0.0004638505  0.0001770763
#>  [4,]  2.186249e-03  0.0023452934  0.0026592979  0.0030176650  0.0035847905
#>  [5,]  1.916256e-03  0.0017770160  0.0016890323  0.0016203792  0.0016177717
#>  [6,] -3.214906e-05 -0.0004002384 -0.0009550839 -0.0013999649 -0.0019778693
#>  [7,] -1.695733e-04 -0.0003582183 -0.0006454688 -0.0008846183 -0.0011994981
#>  [8,]  2.176950e-03  0.0023355905  0.0026296217  0.0029631204  0.0036322711
#>  [9,]  2.252664e-03  0.0023605714  0.0026395544  0.0029198901  0.0034418381
#> [10,]  1.120666e-03  0.0011739427  0.0012444014  0.0013012769  0.0015481699
#>                 V37           V38           V39           V4           V40
#>  [1,]  0.0020887434  0.0019024026  1.666483e-03 2.958977e-04  0.0016264437
#>  [2,]  0.0040662178  0.0029154032  2.273593e-03 3.133415e-04  0.0026043770
#>  [3,]  0.0003186462  0.0009280050  1.077511e-03 2.633917e-04  0.0007562152
#>  [4,]  0.0034688148  0.0023014010  1.699335e-03 2.012096e-04  0.0020719617
#>  [5,]  0.0017332068  0.0017584944  1.606906e-03 3.064674e-04  0.0014922720
#>  [6,] -0.0017085501 -0.0004698748  3.732014e-06 1.375741e-04 -0.0004813784
#>  [7,] -0.0010198264 -0.0003608631 -1.023032e-04 6.609847e-05 -0.0003573688
#>  [8,]  0.0037308194  0.0025692451  1.915976e-03 2.490676e-04  0.0023017738
#>  [9,]  0.0034882593  0.0025309642  1.985983e-03 2.556348e-04  0.0022625991
#> [10,]  0.0016508492  0.0013061082  1.066182e-03 1.365766e-04  0.0011425468
#>                 V41          V42          V43          V44           V45
#>  [1,]  1.609132e-03 0.0014826178 0.0011969959 0.0009685821  8.634943e-04
#>  [2,]  2.046304e-03 0.0016072718 0.0009755572 0.0006286257  1.894037e-04
#>  [3,]  1.114446e-03 0.0012422342 0.0012667228 0.0011541322  1.284115e-03
#>  [4,]  1.566342e-03 0.0011998964 0.0006552570 0.0003556557 -4.003885e-05
#>  [5,]  1.584831e-03 0.0015215053 0.0012987986 0.0010881691  1.051133e-03
#>  [6,]  1.432278e-04 0.0004733893 0.0008314130 0.0009202025  1.251580e-03
#>  [7,] -5.823912e-05 0.0001113903 0.0003231220 0.0003844476  5.365697e-04
#>  [8,]  1.775278e-03 0.0013419998 0.0007279322 0.0004430614  5.334509e-05
#>  [9,]  1.764315e-03 0.0013809977 0.0008339037 0.0005164892  1.418597e-04
#> [10,]  9.570178e-04 0.0007725249 0.0005094950 0.0003807649  2.564647e-04
#>                 V46          V47          V48          V49           V5
#>  [1,]  7.552617e-04 6.032224e-04 0.0004662993 2.539799e-04 3.745189e-04
#>  [2,]  1.350910e-04 2.277467e-04 0.0002233321 1.190822e-04 4.467131e-04
#>  [3,]  1.140121e-03 8.191752e-04 0.0005911909 3.305840e-04 3.155638e-04
#>  [4,] -6.778154e-05 6.541243e-05 0.0001028235 4.367880e-05 2.831268e-04
#>  [5,]  9.262912e-04 7.065998e-04 0.0005328575 2.905294e-04 3.798999e-04
#>  [6,]  1.118574e-03 7.615830e-04 0.0005071426 2.945871e-04 1.477156e-04
#>  [7,]  4.887952e-04 3.422409e-04 0.0002198868 1.327363e-04 7.692385e-05
#>  [8,]  6.780569e-06 1.296667e-04 0.0001525012 8.155956e-05 3.621921e-04
#>  [9,]  8.772879e-05 1.817785e-04 0.0001841228 9.732017e-05 3.693130e-04
#> [10,]  2.117518e-04 2.033006e-04 0.0001692315 9.114872e-05 1.909934e-04
#>                V50          V51          V52          V53          V54
#>  [1,] 1.007814e-04 7.734728e-05 7.344575e-05 5.748679e-05 5.865760e-05
#>  [2,] 6.639838e-05 3.532149e-05 5.985175e-05 5.971651e-05 5.795918e-05
#>  [3,] 1.135998e-04 1.014039e-04 7.713981e-05 5.334655e-05 5.306098e-05
#>  [4,] 4.351868e-05 1.234916e-05 3.478061e-05 3.658891e-05 3.996581e-05
#>  [5,] 1.100231e-04 8.885511e-05 7.998706e-05 6.033131e-05 5.997239e-05
#>  [6,] 8.115629e-05 8.755897e-05 5.291703e-05 2.712452e-05 2.487890e-05
#>  [7,] 3.316191e-05 3.948772e-05 2.212985e-05 1.146248e-05 9.422881e-06
#>  [8,] 4.807101e-05 2.259077e-05 4.748964e-05 4.551517e-05 4.619202e-05
#>  [9,] 5.550006e-05 2.761367e-05 5.193536e-05 5.157240e-05 5.097171e-05
#> [10,] 4.121663e-05 2.656924e-05 3.628605e-05 2.942943e-05 2.932584e-05
#>                 V55          V56           V57          V58          V59
#>  [1,]  5.184175e-05 4.366910e-05  4.905040e-05 4.895417e-05 4.904274e-05
#>  [2,]  7.090516e-05 3.897829e-05  7.396122e-05 4.858627e-05 5.669482e-05
#>  [3,]  3.259901e-05 4.382528e-05  2.531539e-05 4.462016e-05 3.803580e-05
#>  [4,]  5.445019e-05 2.473331e-05  5.773850e-05 3.193545e-05 4.101229e-05
#>  [5,]  4.958566e-05 4.646407e-05  4.581054e-05 5.011148e-05 4.897348e-05
#>  [6,] -5.049787e-06 2.607572e-05 -1.164897e-05 2.111302e-05 9.839318e-06
#>  [7,] -6.006042e-06 1.175744e-05 -8.577019e-06 8.644050e-06 2.683133e-06
#>  [8,]  6.014657e-05 2.727156e-05  6.009476e-05 3.635670e-05 4.647234e-05
#>  [9,]  6.135037e-05 3.324912e-05  6.308533e-05 4.281980e-05 4.832475e-05
#> [10,]  3.266843e-05 1.834369e-05  2.821145e-05 2.284634e-05 2.640049e-05
#>                  V6          V60           V7           V8           V9
#>  [1,]  4.675752e-04 4.068166e-05 5.534030e-04 0.0006636504 0.0009108430
#>  [2,]  7.133503e-04 4.402782e-05 7.943973e-04 0.0007804884 0.0008070525
#>  [3,]  2.776801e-04 3.288637e-05 3.446754e-04 0.0005403343 0.0009360377
#>  [4,]  5.074417e-04 3.379353e-05 5.836988e-04 0.0005272802 0.0004491210
#>  [5,]  4.428682e-04 4.131012e-05 5.248286e-04 0.0006625753 0.0009752742
#>  [6,] -4.128328e-06 1.018845e-05 1.470178e-05 0.0002377983 0.0006742045
#>  [7,] -3.350728e-06 3.126658e-06 1.309685e-06 0.0001074597 0.0003107582
#>  [8,]  6.073867e-04 3.461107e-05 6.765999e-04 0.0006428514 0.0006334325
#>  [9,]  5.947287e-04 3.768862e-05 6.689893e-04 0.0006581978 0.0006778731
#> [10,]  2.862165e-04 2.025710e-05 3.252039e-04 0.0003482822 0.0004334945
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.07790559  0.07479479  0.06866294  0.06053184  0.06861815  0.05888467
#> [2,] -0.07790559 -0.07479479 -0.06866294 -0.06053184 -0.06861815 -0.05888467
#>            [,7]        [,8]        [,9]       [,10]
#> [1,]  0.0466968  0.08588325  0.08383696  0.09273663
#> [2,] -0.0466968 -0.08588325 -0.08383696 -0.09273663
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4878357 0.5071141 0.3906839 0.2504590 0.2962844 0.3077876 0.2767673
#>  [8] 0.6468284 0.3843270 0.8435390
#> 
#> $B[[2]]
#> [1] 0.07160430 0.02294398
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0046613071  0.0069615002  0.0025100366  0.0057853140  0.0043117466
#>  [6] -0.0009008991 -0.0007236555  0.0059350212  0.0059674028  0.0028638404
#> 
#> $vB[[2]]
#> [1]  0.0997513 -0.0997513
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#>  [2,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#>  [3,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#>  [4,] 0.0333 0.0920 0.1475 0.0522 0.1119 0.0970 0.1174 0.1678 0.1642 0.1205
#>  [5,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#>  [6,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#>  [7,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#>  [8,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#>  [9,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [10,] 0.0335 0.2660 0.3188 0.3553 0.3116 0.1965 0.1780 0.2794 0.2870 0.3969
#> [11,] 0.0526 0.1632 0.2544 0.2807 0.3011 0.3361 0.3024 0.2285 0.2910 0.1316
#> [12,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [13,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [14,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [15,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [16,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [17,] 0.0262 0.6194 0.6333 0.7060 0.5544 0.5320 0.6479 0.6931 0.6759 0.7551
#> [18,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [19,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [20,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [21,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [22,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [23,] 0.0715 0.4186 0.4867 0.5249 0.5959 0.6855 0.8573 0.9718 0.8693 0.8711
#> [24,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [25,] 0.0056 0.1859 0.2481 0.2712 0.2934 0.2637 0.1880 0.1405 0.2028 0.2613
#> [26,] 0.0090 0.1895 0.1896 0.2547 0.4073 0.2988 0.2901 0.5326 0.4022 0.1571
#> [27,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [28,] 0.0100 0.2668 0.3376 0.3282 0.2432 0.1268 0.1278 0.4441 0.6795 0.7051
#> [29,] 0.0228 0.3517 0.3330 0.3643 0.4020 0.4731 0.5196 0.6573 0.8426 0.8476
#> [30,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [31,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#> [32,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [33,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [34,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [35,] 0.0162 0.1334 0.2969 0.4754 0.5677 0.5690 0.6421 0.7487 0.8999 1.0000
#> [36,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#> [37,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [38,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [39,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#> [40,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#>  [2,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#>  [3,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#>  [4,] 0.0494 0.0221 0.1544 0.3485 0.6146 0.9146 0.9364 0.8677 0.8772 0.8553
#>  [5,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#>  [6,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#>  [7,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#>  [8,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#>  [9,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [10,] 0.5599 0.0258 0.6936 0.7969 0.7452 0.8203 0.9261 0.8810 0.8814 0.9301
#> [11,] 0.1151 0.0563 0.3404 0.5562 0.6379 0.6553 0.7384 0.6534 0.5423 0.6877
#> [12,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [13,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [14,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [15,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [16,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [17,] 0.8929 0.0582 0.8619 0.7974 0.6737 0.4293 0.3648 0.5331 0.2413 0.5070
#> [18,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [19,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [20,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [21,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [22,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [23,] 0.8954 0.0849 0.9922 0.8980 0.8158 0.8373 0.7541 0.5893 0.5488 0.5643
#> [24,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [25,] 0.2778 0.0267 0.3346 0.3830 0.4003 0.5114 0.6860 0.7490 0.7843 0.9021
#> [26,] 0.3024 0.0062 0.3907 0.3542 0.4438 0.6414 0.4601 0.6009 0.8690 0.8345
#> [27,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [28,] 0.7966 0.0275 0.9401 0.9857 0.8193 0.5789 0.6394 0.7043 0.6875 0.4081
#> [29,] 0.8344 0.0853 0.8453 0.7999 0.8537 0.9642 1.0000 0.9357 0.9409 0.9070
#> [30,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [31,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#> [32,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [33,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [34,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [35,] 0.9690 0.0253 0.9032 0.7685 0.6998 0.6644 0.5964 0.3711 0.0921 0.0481
#> [36,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#> [37,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [38,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [39,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#> [40,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#>  [2,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#>  [3,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#>  [4,] 0.8833 1.0000 0.0270 0.8296 0.6601 0.5499 0.5716 0.6859 0.6825 0.5142
#>  [5,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#>  [6,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#>  [7,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#>  [8,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#>  [9,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [10,] 0.9955 0.8576 0.0398 0.6069 0.3934 0.2464 0.1645 0.1140 0.0956 0.0080
#> [11,] 0.7325 0.7726 0.1219 0.8229 0.8787 0.9108 0.6705 0.6092 0.7505 0.4775
#> [12,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [13,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [14,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [15,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [16,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [17,] 0.8533 0.6036 0.1099 0.8514 0.8512 0.5045 0.1862 0.2709 0.4232 0.3043
#> [18,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [19,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [20,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [21,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [22,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [23,] 0.5406 0.4783 0.0587 0.4439 0.3698 0.2574 0.1478 0.1743 0.1229 0.1588
#> [24,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [25,] 1.0000 0.8888 0.0221 0.6511 0.6083 0.4463 0.2948 0.1729 0.1488 0.0801
#> [26,] 0.7669 0.5081 0.0253 0.4620 0.5380 0.5375 0.3844 0.3601 0.7402 0.7761
#> [27,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [28,] 0.1811 0.2064 0.0190 0.3917 0.3791 0.2042 0.2227 0.3341 0.3984 0.5077
#> [29,] 0.7104 0.6320 0.1000 0.5667 0.3501 0.2447 0.1698 0.3290 0.3674 0.2331
#> [30,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [31,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#> [32,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [33,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [34,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [35,] 0.0876 0.1040 0.0262 0.1714 0.3264 0.4612 0.3939 0.5050 0.4833 0.3511
#> [36,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#> [37,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [38,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [39,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#> [40,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#>  [2,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#>  [3,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#>  [4,] 0.2750 0.1358 0.1551 0.0481 0.2646 0.1994 0.1883 0.2746 0.1651 0.0575
#>  [5,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#>  [6,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#>  [7,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#>  [8,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#>  [9,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [10,] 0.0702 0.0936 0.0894 0.0570 0.1127 0.0873 0.1020 0.1964 0.2256 0.1814
#> [11,] 0.1666 0.3749 0.3776 0.1206 0.2106 0.5886 0.5628 0.2577 0.5245 0.6149
#> [12,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [13,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [14,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [15,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [16,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [17,] 0.6116 0.6756 0.5375 0.1083 0.4719 0.4647 0.2587 0.2129 0.2222 0.2111
#> [18,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [19,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [20,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [21,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [22,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [23,] 0.1803 0.1436 0.1667 0.0218 0.2630 0.2234 0.1239 0.0869 0.2092 0.1499
#> [24,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [25,] 0.1770 0.1382 0.2404 0.0561 0.2046 0.1970 0.2778 0.1377 0.0685 0.0664
#> [26,] 0.3858 0.0667 0.3684 0.0489 0.6114 0.3510 0.2312 0.2195 0.3051 0.1937
#> [27,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [28,] 0.5534 0.3352 0.2723 0.0371 0.2278 0.2044 0.1986 0.0835 0.0908 0.1380
#> [29,] 0.2413 0.2556 0.1892 0.0428 0.1940 0.3074 0.2785 0.0308 0.1238 0.1854
#> [30,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [31,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#> [32,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [33,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [34,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [35,] 0.2319 0.4029 0.3676 0.0386 0.1510 0.0745 0.1395 0.1552 0.0377 0.0636
#> [36,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#> [37,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [38,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [39,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#> [40,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#>  [2,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#>  [3,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#>  [4,] 0.0695 0.0598 0.0456 0.0021 0.0679 0.0068 0.0036 0.0022 0.0032 0.0060
#>  [5,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#>  [6,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#>  [7,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#>  [8,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#>  [9,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [10,] 0.2012 0.1688 0.1037 0.0501 0.0529 0.0136 0.0130 0.0120 0.0039 0.0053
#> [11,] 0.5123 0.3385 0.1499 0.0546 0.0246 0.0270 0.0380 0.0339 0.0149 0.0335
#> [12,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [13,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [14,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [15,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [16,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [17,] 0.0176 0.1348 0.0744 0.0130 0.0974 0.0106 0.0033 0.0232 0.0166 0.0095
#> [18,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [19,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [20,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [21,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [22,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [23,] 0.0676 0.0899 0.0927 0.0658 0.0862 0.0086 0.0216 0.0153 0.0121 0.0096
#> [24,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [25,] 0.1665 0.1807 0.1245 0.0516 0.0936 0.0044 0.0185 0.0072 0.0055 0.0074
#> [26,] 0.1570 0.0479 0.0538 0.0146 0.1197 0.0068 0.0187 0.0059 0.0095 0.0194
#> [27,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [28,] 0.1948 0.1211 0.0843 0.0589 0.0416 0.0247 0.0118 0.0088 0.0104 0.0036
#> [29,] 0.1753 0.1079 0.0728 0.0242 0.1117 0.0191 0.0159 0.0172 0.0191 0.0260
#> [30,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [31,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#> [32,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [33,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [34,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [35,] 0.0443 0.0264 0.0223 0.0187 0.0645 0.0077 0.0137 0.0071 0.0082 0.0232
#> [36,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#> [37,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [38,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [39,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#> [40,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#>  [2,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#>  [3,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#>  [4,] 0.0054 0.0063 0.0143 0.0132 0.0051 0.0981 0.0041 0.0843 0.1172 0.0759
#>  [5,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#>  [6,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#>  [7,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#>  [8,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#>  [9,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [10,] 0.0062 0.0046 0.0045 0.0022 0.0005 0.1091 0.0031 0.1709 0.1684 0.1865
#> [11,] 0.0376 0.0174 0.0132 0.0103 0.0364 0.1022 0.0208 0.0539 0.0439 0.2291
#> [12,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [13,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [14,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [15,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [16,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [17,] 0.0180 0.0244 0.0316 0.0164 0.0095 0.2280 0.0078 0.2431 0.3771 0.5598
#> [18,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [19,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [20,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [21,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [22,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [23,] 0.0196 0.0042 0.0066 0.0099 0.0083 0.1801 0.0124 0.1916 0.1896 0.2960
#> [24,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [25,] 0.0068 0.0084 0.0037 0.0024 0.0034 0.1146 0.0007 0.0706 0.0996 0.1673
#> [26,] 0.0080 0.0152 0.0158 0.0053 0.0189 0.1589 0.0102 0.1392 0.0987 0.0955
#> [27,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [28,] 0.0088 0.0047 0.0117 0.0020 0.0091 0.0201 0.0058 0.0314 0.0651 0.1896
#> [29,] 0.0140 0.0125 0.0116 0.0093 0.0012 0.1651 0.0036 0.1597 0.2116 0.3295
#> [30,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [31,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#> [32,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [33,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [34,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [35,] 0.0198 0.0074 0.0035 0.0100 0.0048 0.0472 0.0019 0.1056 0.1388 0.0598
#> [36,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#> [37,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [38,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [39,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#> [40,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7758377 0.8060306 0.6759826 0.6161025 0.7638422 0.5575555 0.4209571
#>  [2,] 0.7067218 0.6125111 0.5154179 0.4030300 0.5947311 0.3954407 0.2442306
#>  [3,] 0.7561720 0.7534953 0.6251332 0.5240384 0.7102170 0.4760805 0.3306662
#>  [4,] 0.7737124 0.7586975 0.6513464 0.5662926 0.7121805 0.5711773 0.3764946
#>  [5,] 0.6741495 0.7286448 0.5475286 0.5023734 0.6542459 0.4008220 0.3558328
#>  [6,] 0.7191253 0.7647400 0.6072712 0.5229922 0.7457133 0.4100465 0.3145917
#>  [7,] 0.7776973 0.8339402 0.6546484 0.5952817 0.7899890 0.4484277 0.3492668
#>  [8,] 0.7688860 0.8133506 0.6264999 0.6063889 0.7732255 0.4804416 0.4364412
#>  [9,] 0.7604989 0.7873822 0.6664761 0.5982325 0.7335092 0.5441216 0.4027116
#> [10,] 0.7629763 0.8184235 0.6445529 0.6063725 0.7879617 0.5173361 0.3924518
#> [11,] 0.6755966 0.5928765 0.4913833 0.3722297 0.5475356 0.3426039 0.1939457
#> [12,] 0.6833020 0.6000163 0.5507198 0.3906882 0.5325253 0.4187212 0.2321917
#> [13,] 0.6586779 0.5445558 0.5217798 0.3386404 0.4747619 0.3707216 0.1911796
#> [14,] 0.6798756 0.6201475 0.5816123 0.4638166 0.5395693 0.4921481 0.3600005
#> [15,] 0.7425349 0.7527202 0.6254287 0.5370587 0.7060235 0.4727963 0.3286913
#> [16,] 0.7389156 0.7692848 0.6730355 0.5943618 0.7296755 0.5486279 0.4340644
#> [17,] 0.6894173 0.6377483 0.4561507 0.3585136 0.5789370 0.2602226 0.1920776
#> [18,] 0.7855815 0.8117409 0.5956419 0.5608521 0.7648573 0.4305365 0.3554899
#> [19,] 0.7337498 0.7346245 0.6620707 0.5451453 0.6746185 0.4984078 0.3819141
#> [20,] 0.7321126 0.7728686 0.6099582 0.5327520 0.7487035 0.4364406 0.3298349
#> [21,] 0.7648689 0.8180238 0.6418945 0.5974787 0.7913944 0.5049430 0.3732718
#> [22,] 0.7062352 0.7278648 0.6313541 0.5090788 0.6786950 0.4529164 0.3484904
#> [23,] 0.7425662 0.7808429 0.5248035 0.5230090 0.7442142 0.3233270 0.2968236
#> [24,] 0.7008920 0.7168729 0.5367095 0.4857232 0.6490873 0.3089882 0.2757240
#> [25,] 0.7296736 0.7509909 0.6349411 0.5721010 0.6971235 0.5281291 0.3820586
#> [26,] 0.7103667 0.6850461 0.5864766 0.4685940 0.6422620 0.4262742 0.2957855
#> [27,] 0.7032221 0.7024481 0.5557264 0.4499634 0.6553981 0.3881444 0.2382140
#> [28,] 0.7504191 0.7554369 0.6043406 0.5301287 0.7273312 0.4138771 0.3480498
#> [29,] 0.7744054 0.8161926 0.6041628 0.5598019 0.7804355 0.3822760 0.3151708
#> [30,] 0.7707460 0.8169987 0.6043489 0.5601153 0.7679335 0.3963930 0.3318823
#> [31,] 0.7032552 0.6386490 0.5021256 0.3685381 0.6023181 0.3416784 0.2141144
#> [32,] 0.7635579 0.7981778 0.6748993 0.6104721 0.7460179 0.5487854 0.4115424
#> [33,] 0.7473880 0.7304278 0.6424747 0.5716200 0.6775681 0.5781896 0.4130362
#> [34,] 0.7274575 0.7477848 0.5805617 0.4815304 0.7095568 0.4075957 0.2570952
#> [35,] 0.7335819 0.7310497 0.5118958 0.4893981 0.6857887 0.3521900 0.3178052
#> [36,] 0.7064180 0.7009800 0.5670723 0.4669487 0.6234616 0.3340991 0.2817440
#> [37,] 0.7704940 0.7987267 0.6730259 0.6067782 0.7463758 0.5493315 0.3966507
#> [38,] 0.6425176 0.6400639 0.4419238 0.4244788 0.5917476 0.2836365 0.2895375
#> [39,] 0.7462204 0.7580144 0.5829513 0.5028916 0.7185212 0.3799107 0.2965903
#> [40,] 0.7643250 0.8184181 0.6584768 0.6175048 0.7854026 0.5418194 0.4087968
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8960604 0.8383153 0.9564710
#>  [2,] 0.8007769 0.7584791 0.9242356
#>  [3,] 0.8591814 0.8272268 0.9464654
#>  [4,] 0.8786949 0.8256208 0.9577190
#>  [5,] 0.7850140 0.7232517 0.8793324
#>  [6,] 0.8522450 0.8165227 0.9402378
#>  [7,] 0.8945374 0.8633484 0.9580029
#>  [8,] 0.8519346 0.8582498 0.9311540
#>  [9,] 0.8836742 0.8158337 0.9471976
#> [10,] 0.8892287 0.8406960 0.9523121
#> [11,] 0.7899905 0.7065286 0.9247430
#> [12,] 0.8127501 0.6942743 0.9338216
#> [13,] 0.7939932 0.6591176 0.9233561
#> [14,] 0.7781037 0.6759859 0.8581042
#> [15,] 0.8587644 0.8031529 0.9443247
#> [16,] 0.8521055 0.8084997 0.9319291
#> [17,] 0.7972194 0.7216018 0.9063231
#> [18,] 0.8790579 0.8558977 0.9456410
#> [19,] 0.8436688 0.7951337 0.9292712
#> [20,] 0.8644619 0.8165795 0.9446944
#> [21,] 0.8907977 0.8505855 0.9570939
#> [22,] 0.8339531 0.7829671 0.9136508
#> [23,] 0.8299900 0.8169538 0.9282797
#> [24,] 0.7822301 0.7696208 0.9234469
#> [25,] 0.8607423 0.7696611 0.9331004
#> [26,] 0.8174960 0.7602767 0.9238181
#> [27,] 0.8488978 0.7567243 0.9348560
#> [28,] 0.8489324 0.8365496 0.9294534
#> [29,] 0.8722060 0.8622772 0.9563763
#> [30,] 0.8710643 0.8555774 0.9478345
#> [31,] 0.8051479 0.7644727 0.9142614
#> [32,] 0.8861272 0.8250560 0.9503051
#> [33,] 0.8535900 0.7802982 0.9380764
#> [34,] 0.8706426 0.7973492 0.9423615
#> [35,] 0.7960500 0.8075705 0.8939442
#> [36,] 0.7891640 0.7747174 0.9036746
#> [37,] 0.8959410 0.8298777 0.9549118
#> [38,] 0.7033792 0.6738922 0.8185768
#> [39,] 0.8400170 0.8376054 0.9372597
#> [40,] 0.8943674 0.8418502 0.9546764
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.7378735 0.2621265
#>  [2,] 0.7210412 0.2789588
#>  [3,] 0.7337728 0.2662272
#>  [4,] 0.7336008 0.2663992
#>  [5,] 0.7220636 0.2779364
#>  [6,] 0.7348383 0.2651617
#>  [7,] 0.7432031 0.2567969
#>  [8,] 0.7373183 0.2626817
#>  [9,] 0.7350757 0.2649243
#> [10,] 0.7392305 0.2607695
#> [11,] 0.7182226 0.2817774
#> [12,] 0.7172256 0.2827744
#> [13,] 0.7120083 0.2879917
#> [14,] 0.7103417 0.2896583
#> [15,] 0.7329281 0.2670719
#> [16,] 0.7309213 0.2690787
#> [17,] 0.7224687 0.2775313
#> [18,] 0.7405825 0.2594175
#> [19,] 0.7287962 0.2712038
#> [20,] 0.7355145 0.2644855
#> [21,] 0.7402640 0.2597360
#> [22,] 0.7274396 0.2725604
#> [23,] 0.7369335 0.2630665
#> [24,] 0.7284245 0.2715755
#> [25,] 0.7297203 0.2702797
#> [26,] 0.7249170 0.2750830
#> [27,] 0.7289365 0.2710635
#> [28,] 0.7345532 0.2654468
#> [29,] 0.7426628 0.2573372
#> [30,] 0.7413395 0.2586605
#> [31,] 0.7230217 0.2769783
#> [32,] 0.7362380 0.2637620
#> [33,] 0.7275824 0.2724176
#> [34,] 0.7342869 0.2657131
#> [35,] 0.7292997 0.2707003
#> [36,] 0.7260825 0.2739175
#> [37,] 0.7373256 0.2626744
#> [38,] 0.7114428 0.2885572
#> [39,] 0.7354067 0.2645933
#> [40,] 0.7390108 0.2609892
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]         [,3]         [,4]       [,5]         [,6]
#>  [1,] 1.2415728 1.4244210  0.735370333  0.473037964  1.1738608  0.231247079
#>  [2,] 0.8795156 0.4578797  0.061691099 -0.392855885  0.3835587 -0.424498822
#>  [3,] 1.1318059 1.1173419  0.511393990  0.096227702  0.8964383 -0.095751007
#>  [4,] 1.2293937 1.1455516  0.624962757  0.266740712  0.9059979  0.286656100
#>  [5,] 0.7270133 0.9877575  0.190690021  0.009493543  0.6377555 -0.402041359
#>  [6,] 0.9401269 1.1788447  0.435856091  0.092033695  1.0758791 -0.363773252
#>  [7,] 1.2522976 1.6138139  0.639535075  0.385843264  1.3248589 -0.207025324
#>  [8,] 1.2020315 1.4719301  0.517230337  0.432158048  1.2266144 -0.078273723
#>  [9,] 1.1554166 1.3092177  0.692289934  0.398105789  1.0125003  0.176946581
#> [10,] 1.1690668 1.5057026  0.595180668  0.432089644  1.3126828  0.069372069
#> [11,] 0.7336083 0.3758695 -0.034470055 -0.522663305  0.1907186 -0.651711702
#> [12,] 0.7689884 0.4055332  0.203579400 -0.444420347  0.1302852 -0.328025395
#> [13,] 0.6574081 0.1786971  0.087174376 -0.669359145 -0.1010381 -0.529122404
#> [14,] 0.7532002 0.4901744  0.329395713 -0.144986929  0.1586090 -0.031410363
#> [15,] 1.0591859 1.1131733  0.512655164  0.148507015  0.8761486 -0.108922300
#> [16,] 1.0403398 1.2042773  0.721946669  0.382026655  0.9929771  0.195128459
#> [17,] 0.7973966 0.5656044 -0.175849088 -0.581821403  0.3184110 -1.044811787
#> [18,] 1.2984947 1.4613618  0.387338805  0.244620906  1.1794966 -0.279662557
#> [19,] 1.0137318 1.0182135  0.672535823  0.181074256  0.7291489 -0.006368937
#> [20,] 1.0053677 1.2245805  0.447136315  0.131195918  1.0917096 -0.255620467
#> [21,] 1.1795612 1.5030156  0.583596509  0.394970621  1.3333513  0.019772595
#> [22,] 0.8771689 0.9838163  0.538030308  0.036319134  0.7477812 -0.188893863
#> [23,] 1.0593496 1.2705849  0.099295644  0.092100921  1.0679888 -0.738523761
#> [24,] 0.8515490 0.9290024  0.147102673 -0.057122813  0.6150298 -0.804853604
#> [25,] 0.9929673 1.1039041  0.553473408  0.290428478  0.8336373  0.112635364
#> [26,] 0.8971655 0.7770599  0.349418817 -0.125789650  0.5851956 -0.297068607
#> [27,] 0.8626889 0.8589829  0.223835371 -0.200818762  0.6428531 -0.455118790
#> [28,] 1.1008485 1.1278230  0.423584043  0.120660973  0.9811243 -0.347960496
#> [29,] 1.2333557 1.4907618  0.422840751  0.240358025  1.2682060 -0.479899029
#> [30,] 1.2125283 1.4961445  0.423618875  0.241630006  1.1966790 -0.420517162
#> [31,] 0.8628474 0.5695053  0.008502501 -0.538493490  0.4151331 -0.655823586
#> [32,] 1.1722854 1.3749441  0.730428445  0.449296939  1.0774856  0.195764426
#> [33,] 1.0847298 0.9967944  0.586121352  0.288463645  0.7426179  0.315346065
#> [34,] 0.9817612 1.0868325  0.325079676 -0.073911872  0.8932327 -0.373913726
#> [35,] 1.0128725 0.9999548  0.047592172 -0.042414079  0.7805039 -0.609426811
#> [36,] 0.8780501 0.8519690  0.269916015 -0.132398185  0.5042671 -0.689703247
#> [37,] 1.2111026 1.3783552  0.721902626  0.433789461  1.0793756  0.197970190
#> [38,] 0.5863079 0.5756417 -0.233358222 -0.304414032  0.3711945 -0.926494348
#> [39,] 1.0785546 1.1418242  0.334900531  0.011566658  0.9371380 -0.489927438
#> [40,] 1.1765392 1.5056664  0.656513597  0.478970530  1.2974328  0.167669218
#>             [,7]      [,8]      [,9]    [,10]
#>  [1,] -0.3188457 2.1541979 1.6457462 3.089823
#>  [2,] -1.1296236 1.3911572 1.1443590 2.501339
#>  [3,] -0.7051735 1.8085072 1.5660993 2.872406
#>  [4,] -0.5044539 1.9801291 1.5549035 3.120217
#>  [5,] -0.5934975 1.2951288 0.9606489 1.986124
#>  [6,] -0.7787392 1.7523188 1.4929635 2.755760
#>  [7,] -0.6222636 2.1379506 1.8433834 3.127251
#>  [8,] -0.2556182 1.7498554 1.8008289 2.604553
#>  [9,] -0.3941796 2.0276939 1.4883715 2.886951
#> [10,] -0.4370178 2.0828869 1.6634160 2.994214
#> [11,] -1.4245731 1.3248683 0.8785833 2.508607
#> [12,] -1.1959765 1.4679797 0.8201790 2.646931
#> [13,] -1.4423638 1.3491657 0.6593642 2.488845
#> [14,] -0.5753620 1.2546496 0.7353852 1.799632
#> [15,] -0.7141099 1.8050651 1.4061181 2.830934
#> [16,] -0.2652872 1.7512115 1.4402910 2.616707
#> [17,] -1.4365664 1.3690050 0.9524211 2.269544
#> [18,] -0.5949935 1.9835389 1.7816276 2.856253
#> [19,] -0.4814317 1.6857829 1.3561526 2.575548
#> [20,] -0.7089321 1.8528545 1.4933429 2.837988
#> [21,] -0.5182062 2.0989151 1.7392003 3.104887
#> [22,] -0.6256815 1.6139068 1.2830420 2.359049
#> [23,] -0.8624698 1.5855566 1.4958442 2.560559
#> [24,] -0.9657722 1.2787100 1.2061715 2.490128
#> [25,] -0.4808198 1.8214689 1.2063987 2.635319
#> [26,] -0.8674485 1.4994738 1.1541973 2.495391
#> [27,] -1.1624960 1.7259825 1.1348038 2.663792
#> [28,] -0.6276224 1.7262519 1.6327764 2.578323
#> [29,] -0.7760546 1.9206061 1.8343338 3.087550
#> [30,] -0.6996841 1.9104022 1.7790329 2.899758
#> [31,] -1.3003009 1.4187854 1.1773593 2.366814
#> [32,] -0.3575928 2.0517785 1.5509851 2.950881
#> [33,] -0.3514278 1.7630403 1.2674049 2.717929
#> [34,] -1.0611212 1.9066528 1.3698084 2.794198
#> [35,] -0.7638765 1.3617872 1.4343008 2.131678
#> [36,] -0.9358271 1.3198934 1.2351425 2.238737
#> [37,] -0.4194401 2.1529165 1.5847608 3.052998
#> [38,] -0.8976314 0.8634418 0.7258422 1.506735
#> [39,] -0.8635878 1.6583548 1.6405177 2.703956
#> [40,] -0.3689415 2.1361492 1.6720591 3.047544
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.6059892 -0.4289558
#>  [2,] 0.5604972 -0.3891350
#>  [3,] 0.5959122 -0.4179370
#>  [4,] 0.5918603 -0.4211086
#>  [5,] 0.5686075 -0.3861133
#>  [6,] 0.6038057 -0.4155048
#>  [7,] 0.6258642 -0.4368196
#>  [8,] 0.6089579 -0.4231188
#>  [9,] 0.5976449 -0.4228844
#> [10,] 0.6137580 -0.4282150
#> [11,] 0.5542391 -0.3814232
#> [12,] 0.5472011 -0.3835398
#> [13,] 0.5325446 -0.3726133
#> [14,] 0.5248727 -0.3721713
#> [15,] 0.5936149 -0.4159147
#> [16,] 0.5855857 -0.4137160
#> [17,] 0.5713855 -0.3853549
#> [18,] 0.6200402 -0.4289579
#> [19,] 0.5785101 -0.4100136
#> [20,] 0.6051221 -0.4176615
#> [21,] 0.6171458 -0.4301953
#> [22,] 0.5781136 -0.4035573
#> [23,] 0.6144978 -0.4155928
#> [24,] 0.5868904 -0.3997536
#> [25,] 0.5829809 -0.4102230
#> [26,] 0.5709864 -0.3979977
#> [27,] 0.5859701 -0.4032634
#> [28,] 0.6005984 -0.4172495
#> [29,] 0.6273336 -0.4325210
#> [30,] 0.6231768 -0.4297656
#> [31,] 0.5695002 -0.3900000
#> [32,] 0.6010914 -0.4254149
#> [33,] 0.5739621 -0.4084289
#> [34,] 0.6017386 -0.4147441
#> [35,] 0.5896164 -0.4014562
#> [36,] 0.5766428 -0.3981941
#> [37,] 0.6039955 -0.4281190
#> [38,] 0.5414481 -0.3609538
#> [39,] 0.6050356 -0.4171939
#> [40,] 0.6119002 -0.4289331
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.2621265 -0.2621265
#>  [2,] -0.7210412  0.7210412
#>  [3,] -0.7337728  0.7337728
#>  [4,] -0.7336008  0.7336008
#>  [5,]  0.2779364 -0.2779364
#>  [6,]  0.2651617 -0.2651617
#>  [7,]  0.2567969 -0.2567969
#>  [8,] -0.7373183  0.7373183
#>  [9,] -0.7350757  0.7350757
#> [10,]  0.2607695 -0.2607695
#> [11,]  0.2817774 -0.2817774
#> [12,]  0.2827744 -0.2827744
#> [13,]  0.2879917 -0.2879917
#> [14,] -0.7103417  0.7103417
#> [15,]  0.2670719 -0.2670719
#> [16,] -0.7309213  0.7309213
#> [17,] -0.7224687  0.7224687
#> [18,] -0.7405825  0.7405825
#> [19,] -0.7287962  0.7287962
#> [20,]  0.2644855 -0.2644855
#> [21,]  0.2597360 -0.2597360
#> [22,] -0.7274396  0.7274396
#> [23,]  0.2630665 -0.2630665
#> [24,] -0.7284245  0.7284245
#> [25,]  0.2702797 -0.2702797
#> [26,] -0.7249170  0.7249170
#> [27,]  0.2710635 -0.2710635
#> [28,] -0.7345532  0.7345532
#> [29,]  0.2573372 -0.2573372
#> [30,]  0.2586605 -0.2586605
#> [31,] -0.7230217  0.7230217
#> [32,] -0.7362380  0.7362380
#> [33,] -0.7275824  0.7275824
#> [34,]  0.2657131 -0.2657131
#> [35,]  0.2707003 -0.2707003
#> [36,] -0.7260825  0.7260825
#> [37,] -0.7373256  0.7373256
#> [38,]  0.2885572 -0.2885572
#> [39,]  0.2645933 -0.2645933
#> [40,]  0.2609892 -0.2609892
#> 
#> $L
#> [1] 0.6924640 0.6918688 0.8675116 0.6929934 0.7346051 0.7852838
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4782609