Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13           V14
#>  [1,] -0.2362704 -0.09304985 -0.08449238 -0.12029030 -0.17181846  1.423655e-02
#>  [2,] -0.3302978 -0.08133494 -0.10098890 -0.15101519 -0.06746764 -4.274049e-02
#>  [3,] -0.2930448 -0.15032272 -0.11857716 -0.16076704 -0.08640778 -6.182156e-02
#>  [4,] -0.2784214 -0.01141467 -0.10364105 -0.04971042 -0.03364409  2.422877e-05
#>  [5,] -0.3183477 -0.12225534 -0.05591933 -0.06826201 -0.11075545 -3.395883e-02
#>  [6,] -0.2468918 -0.17968874  0.02253857 -0.07072267 -0.07794920 -3.329997e-02
#>  [7,] -0.3138281 -0.12785480 -0.19282176 -0.11848328 -0.19222211 -2.029144e-01
#>  [8,] -0.3000696 -0.10902685 -0.08538554 -0.01281532  0.04619795 -7.924325e-02
#>  [9,] -0.2232062 -0.08073458 -0.18325364 -0.09270367 -0.01138551 -1.624315e-02
#> [10,] -0.3265933 -0.02592453 -0.07147521 -0.05394219  0.04359872  2.732951e-03
#>                V15         V16          V17          V18         V19         V2
#>  [1,] -0.173828272 -0.02595860 -0.147395697 -0.014738541 -0.05692032 -0.3207513
#>  [2,] -0.085647887 -0.11790217 -0.085350054 -0.053539681 -0.08894849 -0.2730747
#>  [3,] -0.039668562  0.10448298  0.147749853  0.003100791  0.20493354 -0.3287114
#>  [4,]  0.043668624 -0.01733390  0.008893936  0.085568607 -0.04071268 -0.1738583
#>  [5,]  0.035144209  0.12668839  0.099250245  0.032548877  0.09578951 -0.3543726
#>  [6,]  0.031785861  0.10291855  0.056568653  0.128986794  0.17858372 -0.3238873
#>  [7,] -0.139709191 -0.08782703 -0.186741262 -0.147846368 -0.15988445 -0.1681555
#>  [8,]  0.012071518 -0.02643022  0.051807741  0.036673072 -0.06246922 -0.2714602
#>  [9,] -0.073500896 -0.01572434  0.151754446  0.112874309  0.11641974 -0.3031631
#> [10,]  0.004785628 -0.02355234  0.024814243  0.051257811  0.15253316 -0.3204790
#>               V20         V21         V22         V23        V24        V25
#>  [1,] -0.06728614  0.04245894 0.009950718 0.006573503 0.03650795 0.07962777
#>  [2,] -0.01299517  0.04410217 0.019266169 0.080969120 0.08790561 0.08125251
#>  [3,]  0.11363904  0.09516320 0.064711874 0.187236432 0.20233936 0.21013823
#>  [4,]  0.01888279  0.07648056 0.018243030 0.068984944 0.07318140 0.09891108
#>  [5,]  0.24522452  0.24056493 0.167632478 0.296883743 0.22008319 0.17958169
#>  [6,]  0.03911721  0.14406704 0.252618470 0.150642310 0.15665426 0.19974966
#>  [7,] -0.03721577 -0.06921848 0.038027231 0.066594598 0.17442412 0.18637309
#>  [8,] -0.04114464 -0.03594169 0.167547762 0.043268785 0.05382156 0.11290054
#>  [9,]  0.15913531  0.15099706 0.157748594 0.050079112 0.10927582 0.17319794
#> [10,]  0.10455683  0.10554295 0.072347637 0.097629176 0.25257424 0.26302856
#>              V26           V27         V28          V29         V3         V30
#>  [1,] 0.06946996  1.953002e-01  0.15876428  0.141477110 -0.3480845  0.07027140
#>  [2,] 0.19841652  1.919222e-01  0.16020168  0.071666158 -0.3712207  0.13874301
#>  [3,] 0.14155180  1.263083e-01  0.10295021  0.108363232 -0.2396490 -0.07651562
#>  [4,] 0.04297363 -9.694512e-05 -0.04738779 -0.098223738 -0.1883472 -0.07292567
#>  [5,] 0.22399857  2.608107e-01  0.28043438  0.156054269 -0.3615767  0.02453382
#>  [6,] 0.21473569  1.876413e-01  0.17558902  0.213211592 -0.2986488  0.16811413
#>  [7,] 0.16655708  1.702756e-01  0.13779106  0.063846828 -0.1429699  0.06926669
#>  [8,] 0.21816096  1.274579e-01  0.14503653 -0.007197983 -0.2966579  0.02888387
#>  [9,] 0.15970370  9.307760e-02  0.11757706  0.055449782 -0.2383043 -0.05637812
#> [10,] 0.26246373  2.195578e-01  0.15950025  0.144145034 -0.2841474  0.14004201
#>               V31         V32          V33          V34          V35
#>  [1,]  0.02609248 -0.02360913 -0.056939871 -0.044909858  0.030046888
#>  [2,]  0.00213343 -0.02390534  0.038593250 -0.041509262 -0.086950878
#>  [3,]  0.03232359  0.05269806 -0.019218082 -0.001748513  0.017129295
#>  [4,] -0.04913963 -0.04297019  0.014567217  0.005830173  0.034249815
#>  [5,]  0.14684315  0.04345166 -0.019369843  0.079748812  0.036756051
#>  [6,]  0.09189540  0.16485365 -0.020356537 -0.022914035  0.019359465
#>  [7,]  0.06258178 -0.08170579  0.058873988 -0.108610505 -0.003305214
#>  [8,] -0.02975020  0.06971770  0.041220644 -0.010750941  0.068170808
#>  [9,] -0.09745975  0.02769483 -0.009826667 -0.025531092 -0.130039082
#> [10,]  0.02972399 -0.00667511 -0.121532184 -0.148803356 -0.167543968
#>               V36          V37          V38         V39         V4         V40
#>  [1,] -0.03415803 -0.000368026 -0.007981818 -0.12798571 -0.3116911 -0.08778970
#>  [2,] -0.18934498 -0.042832112 -0.201824105 -0.22171212 -0.3082034 -0.06375221
#>  [3,] -0.03058875  0.070356105 -0.006658302 -0.02036544 -0.2802543 -0.02334208
#>  [4,]  0.06233275  0.116134494 -0.059777901 -0.10977746 -0.2341595  0.04997661
#>  [5,]  0.05305197 -0.017639047 -0.068337598  0.05659888 -0.3625091 -0.04944419
#>  [6,]  0.02792539  0.093800302 -0.092309289 -0.09580453 -0.1844922 -0.06889656
#>  [7,] -0.03344123 -0.097790161 -0.144024829 -0.04017446 -0.1303534 -0.01716505
#>  [8,]  0.07994054 -0.065370650  0.032756587 -0.03861245 -0.2149694  0.02653750
#>  [9,] -0.01748451  0.038281007 -0.010702023 -0.04814422 -0.3114347 -0.14760893
#> [10,] -0.02033367 -0.118618428 -0.136837843  0.01510547 -0.3058410 -0.11168872
#>               V41          V42          V43         V44         V45         V46
#>  [1,] -0.06535507 -0.200817461 -0.165581730 -0.12443260 -0.11984712 -0.24643514
#>  [2,] -0.06037893 -0.100576447 -0.117579175 -0.02264275 -0.03183492 -0.14378026
#>  [3,] -0.07532220 -0.132789960 -0.114579285 -0.16230587 -0.17769766 -0.12942318
#>  [4,] -0.14917619 -0.181215854 -0.190237609 -0.01971160 -0.08558506 -0.11941513
#>  [5,] -0.04951298  0.010986245 -0.019494788 -0.07330385 -0.18318584 -0.06870128
#>  [6,]  0.03820994 -0.128096250 -0.149001446 -0.17525185 -0.01569701 -0.14804870
#>  [7,] -0.12203475 -0.158832706 -0.073527575 -0.16498925 -0.15161727 -0.15600372
#>  [8,]  0.01420980 -0.120032636 -0.038977171 -0.06386999 -0.20037805 -0.15887953
#>  [9,] -0.16514774 -0.040114974 -0.124959086 -0.17169134 -0.21331929 -0.16446205
#> [10,] -0.11376362 -0.005927514 -0.009741987 -0.07102651 -0.05567490 -0.15848322
#>               V47        V48        V49         V5        V50        V51
#>  [1,] -0.31379049 -0.2736862 -0.2260671 -0.2007434 -0.3241460 -0.3715314
#>  [2,] -0.15402976 -0.1688902 -0.1876824 -0.2240801 -0.2493044 -0.3634072
#>  [3,] -0.17556485 -0.1775724 -0.3403308 -0.1856078 -0.2794157 -0.3555603
#>  [4,] -0.12152508 -0.2275509 -0.3485999 -0.3211801 -0.3359344 -0.2546823
#>  [5,] -0.08305809 -0.1848241 -0.2230280 -0.2693079 -0.4286702 -0.2627584
#>  [6,] -0.13635652 -0.1366125 -0.1989837 -0.2164998 -0.2841749 -0.3300469
#>  [7,] -0.12183075 -0.2033490 -0.2005431 -0.2018567 -0.3224391 -0.2584492
#>  [8,] -0.14646045 -0.2506300 -0.2962428 -0.1569004 -0.2865450 -0.3100824
#>  [9,] -0.24232398 -0.2330019 -0.2246281 -0.2311530 -0.2400132 -0.2404746
#> [10,] -0.10503346 -0.2944318 -0.3466761 -0.3459582 -0.3648870 -0.3699164
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.4316606 -0.4650254 -0.3000919 -0.2600164 -0.2995500 -0.3460373
#>  [2,] -0.3803333 -0.2414190 -0.4233091 -0.3678237 -0.2333618 -0.3565819
#>  [3,] -0.2422559 -0.2554633 -0.2456953 -0.3465760 -0.4139342 -0.3525250
#>  [4,] -0.3572251 -0.3295657 -0.2587142 -0.2972965 -0.4034512 -0.2273451
#>  [5,] -0.3502031 -0.3519938 -0.2853261 -0.4056205 -0.2830372 -0.3689525
#>  [6,] -0.2876536 -0.2776364 -0.2384201 -0.2748260 -0.3162510 -0.2036505
#>  [7,] -0.2780199 -0.2875960 -0.3176520 -0.3375300 -0.2602997 -0.3791623
#>  [8,] -0.2391601 -0.3205722 -0.3407233 -0.2678399 -0.3713892 -0.3221349
#>  [9,] -0.3103696 -0.4043958 -0.3606666 -0.3437563 -0.3535843 -0.2960912
#> [10,] -0.3739679 -0.3134261 -0.3596378 -0.3732953 -0.4432833 -0.3783007
#>              V58        V59          V6        V60          V7          V8
#>  [1,] -0.4024819 -0.3209287 -0.23955251 -0.4440984 -0.12232932 -0.16211780
#>  [2,] -0.3071726 -0.3967243 -0.19611262 -0.3891651 -0.17465020 -0.20151325
#>  [3,] -0.2379808 -0.3432458 -0.18609006 -0.3098366 -0.20468458 -0.25270566
#>  [4,] -0.2506893 -0.2793412 -0.21805431 -0.2994697 -0.21064587 -0.10203598
#>  [5,] -0.3835157 -0.3469820 -0.11780925 -0.3992491 -0.07337707 -0.09837258
#>  [6,] -0.3863540 -0.2684494 -0.23080410 -0.2084704 -0.15892633 -0.10593536
#>  [7,] -0.3400430 -0.2936821 -0.09861821 -0.1911793 -0.17617596 -0.12297286
#>  [8,] -0.3688571 -0.2102744 -0.10810085 -0.3709024 -0.05213990 -0.11330755
#>  [9,] -0.3510653 -0.3819814 -0.23950179 -0.2218454 -0.24575024 -0.26794666
#> [10,] -0.2871426 -0.3507813 -0.18460099 -0.4263364 -0.11362923 -0.21999262
#>                V9
#>  [1,] -0.08065635
#>  [2,] -0.13042070
#>  [3,] -0.16929235
#>  [4,] -0.15253969
#>  [5,] -0.19884724
#>  [6,] -0.18008098
#>  [7,] -0.09559297
#>  [8,] -0.08987483
#>  [9,] -0.07398905
#> [10,] -0.19979283
#> 
#> $W[[2]]
#>             [,1]        [,2]       [,3]        [,4]        [,5]       [,6]
#> [1,] -0.16450458 -0.02336300 0.06925845 -0.12251797 0.015849776 0.08567176
#> [2,]  0.01751578 -0.04189495 0.05678360  0.01621363 0.007959651 0.01423092
#>            [,7]       [,8]         [,9]        [,10]
#> [1,] 0.02393284 0.06178573 -0.006434481  0.008070708
#> [2,] 0.02018561 0.05192709  0.065066327 -0.008176607
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 3.733830e-04 0.0025884892 0.0030060944 0.0032842911 0.0031218007
#>  [2,] 5.728340e-05 0.0005162671 0.0005229877 0.0006536723 0.0007882368
#>  [3,] 1.074062e-04 0.0008496943 0.0008987452 0.0010450426 0.0011010789
#>  [4,] 3.284449e-04 0.0022538900 0.0026203855 0.0028650984 0.0027196750
#>  [5,] 1.021631e-04 0.0007069242 0.0007362451 0.0008003057 0.0008024253
#>  [6,] 4.787435e-05 0.0004192682 0.0004136344 0.0005155454 0.0005928571
#>  [7,] 4.719056e-05 0.0004554452 0.0004396553 0.0005761919 0.0007413231
#>  [8,] 1.060650e-04 0.0008350846 0.0008973401 0.0010495240 0.0011361027
#>  [9,] 2.442808e-04 0.0017163717 0.0019397164 0.0021437442 0.0020784786
#> [10,] 1.267206e-04 0.0009171287 0.0009748543 0.0010847690 0.0011033242
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0031029555 0.0030824260 0.0035174783 0.0035123456 0.0035565207
#>  [2,] 0.0010408450 0.0011976012 0.0015083301 0.0015535436 0.0015445208
#>  [3,] 0.0012285098 0.0013098616 0.0015521120 0.0015277694 0.0015199377
#>  [4,] 0.0026787363 0.0026503592 0.0030380875 0.0030378363 0.0030923434
#>  [5,] 0.0008497372 0.0009068883 0.0010355121 0.0009674935 0.0009421469
#>  [6,] 0.0007156468 0.0008020114 0.0009649308 0.0009454073 0.0009469023
#>  [7,] 0.0010348586 0.0012121605 0.0015429693 0.0015897513 0.0015726737
#>  [8,] 0.0013292751 0.0014598751 0.0017649593 0.0017795986 0.0017851037
#>  [9,] 0.0020962104 0.0021087441 0.0024336722 0.0024054399 0.0024065750
#> [10,] 0.0012068376 0.0012929850 0.0015156856 0.0014768970 0.0014495151
#>                V19           V2         V20         V21         V22         V23
#>  [1,] 0.0046151088 4.527627e-04 0.005905681 0.007283499 0.007264922 0.006619415
#>  [2,] 0.0015636165 9.944529e-05 0.001872432 0.002360678 0.002454839 0.002444863
#>  [3,] 0.0016917451 1.547855e-04 0.002107626 0.002654372 0.002651078 0.002544045
#>  [4,] 0.0040051113 4.001210e-04 0.005099519 0.006290917 0.006291063 0.005756087
#>  [5,] 0.0011031896 1.290071e-04 0.001454032 0.001847432 0.001763133 0.001598780
#>  [6,] 0.0009717162 7.972071e-05 0.001186881 0.001515543 0.001516730 0.001489815
#>  [7,] 0.0015430060 9.069254e-05 0.001838526 0.002336533 0.002429751 0.002428068
#>  [8,] 0.0019594997 1.544379e-04 0.002411204 0.003015808 0.003064726 0.002949135
#>  [9,] 0.0030077726 3.090793e-04 0.003833372 0.004760988 0.004723018 0.004361985
#> [10,] 0.0016532017 1.697662e-04 0.002109096 0.002658832 0.002589972 0.002394165
#>               V24         V25         V26         V27         V28         V29
#>  [1,] 0.006933588 0.007203682 0.007402857 0.007786717 0.007256462 0.005816585
#>  [2,] 0.002779757 0.003066141 0.003211653 0.003282587 0.003147307 0.002829985
#>  [3,] 0.002882569 0.003148608 0.003262655 0.003378289 0.003293658 0.002903644
#>  [4,] 0.006038648 0.006271874 0.006431558 0.006759624 0.006321305 0.005070083
#>  [5,] 0.001809369 0.001972185 0.002023740 0.002102361 0.002080602 0.001809543
#>  [6,] 0.001750147 0.001930925 0.001987477 0.002034202 0.002019992 0.001855933
#>  [7,] 0.002791329 0.003106112 0.003269136 0.003335656 0.003197478 0.002883604
#>  [8,] 0.003282837 0.003558995 0.003680980 0.003784328 0.003620410 0.003172687
#>  [9,] 0.004702052 0.004998443 0.005155807 0.005411498 0.005143190 0.004250750
#> [10,] 0.002688815 0.002931717 0.003020687 0.003134314 0.003037641 0.002632357
#>                 V3         V30         V31         V32          V33
#>  [1,] 5.296192e-04 0.005188557 0.003597740 0.003240756 0.0022796714
#>  [2,] 1.110509e-04 0.002449027 0.002110244 0.001773375 0.0014949176
#>  [3,] 1.726720e-04 0.002570406 0.002090761 0.001821891 0.0014709382
#>  [4,] 4.610864e-04 0.004498393 0.003135746 0.002831406 0.0019845936
#>  [5,] 1.384508e-04 0.001659008 0.001281051 0.001154806 0.0009323730
#>  [6,] 8.863579e-05 0.001620051 0.001369556 0.001168727 0.0009947935
#>  [7,] 1.007647e-04 0.002482164 0.002173063 0.001822586 0.0015509178
#>  [8,] 1.762659e-04 0.002777056 0.002266154 0.001938701 0.0015810017
#>  [9,] 3.517915e-04 0.003796790 0.002808570 0.002516995 0.0018552497
#> [10,] 1.903863e-04 0.002374544 0.001836535 0.001612594 0.0012827135
#>                V34          V35          V36          V37          V38
#>  [1,] 0.0018414238 0.0014930195 0.0021163935 0.0026043176 0.0031185440
#>  [2,] 0.0014404100 0.0013901298 0.0016010942 0.0015235727 0.0012120041
#>  [3,] 0.0013668241 0.0012468963 0.0015014744 0.0015292380 0.0014049007
#>  [4,] 0.0015872662 0.0012807438 0.0017938769 0.0022334216 0.0026827948
#>  [5,] 0.0008745795 0.0007778924 0.0009250646 0.0009812776 0.0009847644
#>  [6,] 0.0009817958 0.0009314380 0.0010919862 0.0010477362 0.0008598113
#>  [7,] 0.0015027869 0.0014627358 0.0016795990 0.0015788007 0.0012137559
#>  [8,] 0.0014861482 0.0013774963 0.0016425569 0.0016514779 0.0014717485
#>  [9,] 0.0015766482 0.0013171968 0.0017076650 0.0019910962 0.0022160075
#> [10,] 0.0012159381 0.0011265836 0.0013564046 0.0014249072 0.0013836223
#>                V39           V4          V40          V41          V42
#>  [1,] 0.0030262521 0.0007019363 0.0023207869 0.0020770686 0.0024201957
#>  [2,] 0.0010910196 0.0001620471 0.0010933014 0.0007869777 0.0005699384
#>  [3,] 0.0013304838 0.0002332498 0.0012129811 0.0009698630 0.0008452754
#>  [4,] 0.0026047388 0.0006024996 0.0020003956 0.0017921067 0.0020581918
#>  [5,] 0.0009314363 0.0001881122 0.0007890412 0.0006797266 0.0006236234
#>  [6,] 0.0008108564 0.0001238017 0.0007996394 0.0006160387 0.0004658188
#>  [7,] 0.0010785810 0.0001524147 0.0011159889 0.0007776902 0.0005185175
#>  [8,] 0.0013763295 0.0002424353 0.0012837755 0.0010029777 0.0008658610
#>  [9,] 0.0021330358 0.0004635462 0.0017008955 0.0014928076 0.0015755619
#> [10,] 0.0013030008 0.0002597231 0.0011001711 0.0009220617 0.0008513828
#>                V43          V44          V45          V46          V47
#>  [1,] 0.0029002657 0.0030681230 0.0032888350 0.0024225245 0.0016895137
#>  [2,] 0.0006073321 0.0006290838 0.0005034963 0.0003227812 0.0002959872
#>  [3,] 0.0009538792 0.0010000397 0.0009812734 0.0006952758 0.0005225513
#>  [4,] 0.0024716470 0.0026279057 0.0028120043 0.0020875661 0.0014614792
#>  [5,] 0.0007332878 0.0007797258 0.0008400043 0.0006284398 0.0004410755
#>  [6,] 0.0004804275 0.0004780892 0.0004282899 0.0003014409 0.0002510110
#>  [7,] 0.0005495960 0.0005729582 0.0004210075 0.0002496951 0.0002515248
#>  [8,] 0.0009563712 0.0009833128 0.0009252532 0.0006502021 0.0005161928
#>  [9,] 0.0018702018 0.0020123899 0.0021485900 0.0015814466 0.0011084270
#> [10,] 0.0009768582 0.0010373886 0.0010828552 0.0007945428 0.0005703673
#>                V48          V49           V5          V50          V51
#>  [1,] 0.0014195940 0.0008217106 0.0010700555 2.059106e-04 1.929093e-04
#>  [2,] 0.0002832257 0.0001439110 0.0002936824 5.606230e-05 3.290654e-05
#>  [3,] 0.0004652314 0.0002624279 0.0003742476 7.900164e-05 5.387657e-05
#>  [4,] 0.0012229910 0.0007101549 0.0009177978 1.781389e-04 1.679465e-04
#>  [5,] 0.0003817827 0.0002219005 0.0002736957 6.403161e-05 4.514347e-05
#>  [6,] 0.0002316761 0.0001284918 0.0002090450 4.463998e-05 2.522558e-05
#>  [7,] 0.0002552847 0.0001257876 0.0002874484 5.390411e-05 2.851069e-05
#>  [8,] 0.0004605621 0.0002528161 0.0004029983 8.067307e-05 5.692108e-05
#>  [9,] 0.0009414855 0.0005440831 0.0006951104 1.416459e-04 1.213566e-04
#> [10,] 0.0005007804 0.0002874195 0.0003905654 8.194369e-05 5.988212e-05
#>                V52          V53          V54          V55          V56
#>  [1,] 1.581563e-04 1.253607e-04 1.495572e-04 6.309014e-05 1.083172e-04
#>  [2,] 2.529740e-05 3.990188e-05 3.447011e-05 1.719297e-05 3.047162e-05
#>  [3,] 4.242047e-05 4.713179e-05 4.940090e-05 2.435818e-05 4.145850e-05
#>  [4,] 1.386736e-04 1.074189e-04 1.277987e-04 5.607495e-05 9.381042e-05
#>  [5,] 3.512455e-05 3.422973e-05 3.929368e-05 1.949869e-05 3.320759e-05
#>  [6,] 1.750279e-05 2.696665e-05 2.647052e-05 1.329844e-05 2.424719e-05
#>  [7,] 2.040479e-05 3.959493e-05 3.228202e-05 1.607315e-05 2.941968e-05
#>  [8,] 4.444737e-05 5.180574e-05 5.190935e-05 2.457441e-05 4.289729e-05
#>  [9,] 1.008069e-04 8.388866e-05 9.644161e-05 4.580537e-05 7.392835e-05
#> [10,] 4.740412e-05 4.722683e-05 5.333791e-05 2.594817e-05 4.420928e-05
#>                V57          V58          V59           V6          V60
#>  [1,] 5.818023e-05 1.093093e-04 1.007477e-04 0.0012532887 8.398739e-05
#>  [2,] 2.079817e-05 1.623923e-05 2.357087e-05 0.0004120329 2.121228e-05
#>  [3,] 2.567118e-05 3.042086e-05 3.475385e-05 0.0005051827 3.058598e-05
#>  [4,] 5.074836e-05 9.478683e-05 8.677374e-05 0.0010878918 7.190225e-05
#>  [5,] 1.968091e-05 2.667702e-05 2.891189e-05 0.0003637752 2.611523e-05
#>  [6,] 1.559177e-05 1.379442e-05 1.911528e-05 0.0003005489 1.738894e-05
#>  [7,] 2.034447e-05 1.359157e-05 2.212707e-05 0.0004047401 2.029339e-05
#>  [8,] 2.698206e-05 3.055644e-05 3.575043e-05 0.0005357001 3.088148e-05
#>  [9,] 4.179695e-05 6.833650e-05 6.691976e-05 0.0008571838 5.675317e-05
#> [10,] 2.588499e-05 3.488796e-05 3.790551e-05 0.0005053024 3.411608e-05
#>                 V7           V8           V9
#>  [1,] 0.0011971665 0.0014328869 0.0024034779
#>  [2,] 0.0004231641 0.0004342923 0.0005108326
#>  [3,] 0.0005021033 0.0005777915 0.0008166443
#>  [4,] 0.0010431310 0.0012420647 0.0020790492
#>  [5,] 0.0003563454 0.0004429455 0.0006792624
#>  [6,] 0.0002997137 0.0003360953 0.0004096064
#>  [7,] 0.0004182907 0.0004180698 0.0004606845
#>  [8,] 0.0005325079 0.0005933080 0.0007988596
#>  [9,] 0.0008363786 0.0009910193 0.0016118775
#> [10,] 0.0004940929 0.0005932839 0.0008887167
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]      [,5]        [,6]
#> [1,] -0.03363452 -0.03639175 -0.05733671 -0.03431341 -0.074114 -0.06852307
#> [2,]  0.03363452  0.03639175  0.05733671  0.03431341  0.074114  0.06852307
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.02934707 -0.04934937 -0.04949461 -0.06631208
#> [2,]  0.02934707  0.04934937  0.04949461  0.06631208
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5321221 0.3053855 0.6619223 0.4128977 0.6907175 0.5537796 0.3789839
#>  [8] 0.4969530 0.3847745 0.6499803
#> 
#> $B[[2]]
#> [1]  0.08749228 -0.14701274
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.008265635 0.003908685 0.003991400 0.007153238 0.002535425 0.002560261
#>  [7] 0.003994711 0.004447780 0.005912853 0.003704593
#> 
#> $vB[[2]]
#> [1] -0.08287331  0.08287331
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [2,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [3,] 0.0192 0.1186 0.1237 0.1601 0.3520 0.4479 0.3769 0.5761 0.6426 0.6790
#>  [4,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>  [5,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#>  [6,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#>  [7,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#>  [8,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#>  [9,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [10,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [11,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#> [12,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [13,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [14,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [15,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [16,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [17,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [18,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [19,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [20,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [21,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [22,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [23,] 0.0207 0.1234 0.1796 0.1787 0.1247 0.2577 0.3370 0.3990 0.1647 0.2266
#> [24,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [25,] 0.0530 0.0740 0.1610 0.2226 0.2703 0.3365 0.4266 0.4144 0.5655 0.6921
#> [26,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#> [27,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [28,] 0.0235 0.3674 0.2974 0.0837 0.1912 0.5040 0.6352 0.6804 0.7505 0.6595
#> [29,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [30,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [31,] 0.0195 0.1191 0.1522 0.1322 0.1434 0.1244 0.0653 0.0890 0.1226 0.1846
#> [32,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [33,] 0.0126 0.5920 0.5832 0.5419 0.5472 0.5314 0.4981 0.6985 0.8292 0.7839
#> [34,] 0.0294 0.3473 0.4231 0.5044 0.5237 0.4398 0.3236 0.2956 0.3286 0.3231
#> [35,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [36,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#> [37,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [38,] 0.0366 0.1847 0.2222 0.2648 0.2508 0.2291 0.1555 0.1863 0.2387 0.3345
#> [39,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [40,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [2,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [3,] 0.7157 0.0607 0.5466 0.5399 0.6362 0.7849 0.7756 0.5780 0.4862 0.4181
#>  [4,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>  [5,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#>  [6,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#>  [7,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#>  [8,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#>  [9,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [10,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [11,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#> [12,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [13,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [14,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [15,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [16,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [17,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [18,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [19,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [20,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [21,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [22,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [23,] 0.3219 0.0535 0.5356 0.8159 1.0000 0.8701 0.6889 0.6299 0.5738 0.5707
#> [24,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [25,] 0.8547 0.0885 0.9234 0.9171 1.0000 0.9532 0.9101 0.8337 0.7053 0.6534
#> [26,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#> [27,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [28,] 0.4509 0.0291 0.2964 0.4019 0.6794 0.8297 1.0000 0.8240 0.7115 0.7726
#> [29,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [30,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [31,] 0.3880 0.0142 0.3658 0.2297 0.2610 0.4193 0.5848 0.5643 0.5448 0.4772
#> [32,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [33,] 0.8215 0.0149 0.9363 1.0000 0.9224 0.7839 0.5470 0.4562 0.5922 0.5448
#> [34,] 0.4528 0.0123 0.6339 0.7044 0.8314 0.8449 0.8512 0.9138 0.9985 1.0000
#> [35,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [36,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#> [37,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [38,] 0.5233 0.0421 0.6684 0.7766 0.7928 0.7940 0.9129 0.9498 0.9835 1.0000
#> [39,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [40,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [2,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [3,] 0.2457 0.0716 0.0378 0.0613 0.1816 0.4493 0.5976 0.3785 0.2495 0.5771
#>  [4,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>  [5,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#>  [6,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#>  [7,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#>  [8,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#>  [9,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [10,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [11,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#> [12,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [13,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [14,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [15,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [16,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [17,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [18,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [19,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [20,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [21,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [22,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [23,] 0.5976 0.4301 0.0334 0.2058 0.1000 0.2247 0.2308 0.3977 0.3317 0.1726
#> [24,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [25,] 0.4483 0.2460 0.1997 0.2020 0.1446 0.0994 0.1510 0.2392 0.4434 0.5023
#> [26,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#> [27,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [28,] 0.6124 0.4936 0.0749 0.5648 0.4906 0.1820 0.1811 0.1107 0.4603 0.6650
#> [29,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [30,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [31,] 0.6897 0.9797 0.0181 1.0000 0.9546 0.8835 0.7662 0.6547 0.5447 0.4593
#> [32,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [33,] 0.3971 0.0882 0.0641 0.2385 0.2005 0.0587 0.2544 0.2009 0.0329 0.1547
#> [34,] 0.7544 0.4661 0.0117 0.3924 0.3849 0.4674 0.4245 0.3095 0.0752 0.2885
#> [35,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [36,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#> [37,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [38,] 0.9471 0.8237 0.0504 0.6252 0.4181 0.3209 0.2658 0.2196 0.1588 0.0561
#> [39,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [40,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [2,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [3,] 0.8852 0.8409 0.3570 0.0774 0.3133 0.6096 0.6378 0.2709 0.1419 0.1260
#>  [4,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>  [5,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#>  [6,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#>  [7,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#>  [8,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#>  [9,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [10,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [11,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#> [12,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [13,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [14,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [15,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [16,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [17,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [18,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [19,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [20,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [21,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [22,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [23,] 0.1429 0.2168 0.1967 0.0818 0.2140 0.3674 0.2023 0.0778 0.0925 0.2388
#> [24,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [25,] 0.4441 0.4571 0.3927 0.2604 0.2900 0.3408 0.4990 0.3632 0.1387 0.1800
#> [26,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#> [27,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [28,] 0.6423 0.2166 0.1951 0.0519 0.4947 0.4925 0.4041 0.2402 0.1392 0.1779
#> [29,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [30,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [31,] 0.4679 0.1987 0.0699 0.0406 0.1493 0.1713 0.1654 0.2600 0.3846 0.3754
#> [32,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [33,] 0.1212 0.2446 0.3171 0.1732 0.3195 0.3051 0.0836 0.1266 0.1381 0.1136
#> [34,] 0.4072 0.3170 0.2863 0.0113 0.2634 0.0541 0.1874 0.3459 0.4646 0.4366
#> [35,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [36,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#> [37,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [38,] 0.0948 0.1700 0.1215 0.0250 0.1282 0.0386 0.1329 0.2331 0.2468 0.1960
#> [39,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [40,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [2,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [3,] 0.1288 0.0790 0.0829 0.0520 0.1388 0.0216 0.0360 0.0331 0.0131 0.0120
#>  [4,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>  [5,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#>  [6,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#>  [7,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#>  [8,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#>  [9,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [10,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [11,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#> [12,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [13,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [14,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [15,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [16,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [17,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [18,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [19,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [20,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [21,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [22,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [23,] 0.3400 0.2594 0.1102 0.0911 0.0740 0.0462 0.0171 0.0033 0.0050 0.0190
#> [24,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [25,] 0.1299 0.0523 0.0817 0.0469 0.3225 0.0114 0.0299 0.0244 0.0199 0.0257
#> [26,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#> [27,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [28,] 0.1946 0.1723 0.1522 0.0929 0.0227 0.0179 0.0242 0.0083 0.0037 0.0095
#> [29,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [30,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [31,] 0.2414 0.1077 0.0224 0.0155 0.0391 0.0187 0.0125 0.0028 0.0067 0.0120
#> [32,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [33,] 0.0516 0.0073 0.0278 0.0372 0.2565 0.0121 0.0153 0.0092 0.0035 0.0098
#> [34,] 0.2581 0.1319 0.0505 0.0112 0.0497 0.0059 0.0041 0.0056 0.0104 0.0079
#> [35,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [36,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#> [37,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [38,] 0.1985 0.1570 0.0921 0.0549 0.0596 0.0194 0.0166 0.0132 0.0027 0.0022
#> [39,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [40,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [2,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [3,] 0.0108 0.0024 0.0045 0.0037 0.0112 0.0809 0.0075 0.0568 0.0219 0.1037
#>  [4,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#>  [5,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#>  [6,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#>  [7,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#>  [8,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#>  [9,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [10,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [11,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#> [12,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [13,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [14,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [15,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [16,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [17,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [18,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [19,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [20,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [21,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [22,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [23,] 0.0103 0.0121 0.0042 0.0090 0.0070 0.0324 0.0099 0.0918 0.1070 0.1553
#> [24,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [25,] 0.0082 0.0151 0.0171 0.0146 0.0134 0.2247 0.0056 0.0617 0.2287 0.0950
#> [26,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#> [27,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [28,] 0.0105 0.0030 0.0132 0.0068 0.0108 0.0834 0.0090 0.0677 0.2002 0.2876
#> [29,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [30,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [31,] 0.0012 0.0022 0.0058 0.0042 0.0067 0.0249 0.0012 0.0892 0.0973 0.0840
#> [32,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [33,] 0.0121 0.0006 0.0181 0.0094 0.0116 0.2559 0.0063 0.2947 0.4110 0.4983
#> [34,] 0.0014 0.0054 0.0015 0.0006 0.0081 0.0998 0.0043 0.1326 0.1117 0.2984
#> [35,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [36,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#> [37,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [38,] 0.0059 0.0016 0.0025 0.0017 0.0027 0.0252 0.0027 0.0958 0.0991 0.1419
#> [39,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [40,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6698817 0.6075939 0.8017110 0.5724352 0.9119451 0.8802007 0.6100545
#>  [2,] 0.6595506 0.6465689 0.7678577 0.5786394 0.8664450 0.8147077 0.6616642
#>  [3,] 0.4646547 0.4025846 0.7678029 0.5641551 0.8556832 0.7992873 0.3783697
#>  [4,] 0.5421330 0.5149824 0.8158396 0.5644786 0.9243025 0.8795795 0.4491937
#>  [5,] 0.4289893 0.3279266 0.5969132 0.4884234 0.7696589 0.6850637 0.3329875
#>  [6,] 0.6081312 0.5813547 0.7597500 0.5203567 0.8824768 0.8467567 0.5597391
#>  [7,] 0.5162471 0.5007893 0.7325387 0.5307524 0.8775607 0.8103221 0.5075307
#>  [8,] 0.5520525 0.5114232 0.7983167 0.6081948 0.8963576 0.8480390 0.4609648
#>  [9,] 0.6367724 0.6297986 0.8378328 0.5679580 0.9385951 0.9047250 0.5904164
#> [10,] 0.6564376 0.6404736 0.8059118 0.5626260 0.9116650 0.8730664 0.6256894
#> [11,] 0.6877376 0.6763892 0.8016040 0.5538073 0.9099095 0.8829101 0.6834945
#> [12,] 0.5648665 0.5341474 0.8097792 0.5801283 0.9014442 0.8551500 0.4967426
#> [13,] 0.6717687 0.6474423 0.7634646 0.5755739 0.8633983 0.8225669 0.6566337
#> [14,] 0.5419943 0.5385371 0.8077393 0.5590316 0.9175792 0.8886969 0.4399458
#> [15,] 0.5992188 0.5974878 0.8441324 0.5739890 0.9403172 0.9058044 0.5385032
#> [16,] 0.5682892 0.5540983 0.8310186 0.5731117 0.9301980 0.8831992 0.5434899
#> [17,] 0.4429043 0.4873421 0.6253363 0.3991618 0.8338941 0.7728556 0.4281731
#> [18,] 0.5165060 0.4719014 0.6769424 0.4850109 0.8551317 0.7970973 0.4595572
#> [19,] 0.5434838 0.4823650 0.7884059 0.5568871 0.9036262 0.8520928 0.4945655
#> [20,] 0.4606106 0.4058534 0.6728197 0.4876380 0.8078887 0.7277820 0.4392561
#> [21,] 0.6604501 0.6379743 0.7977932 0.5573421 0.9082410 0.8694410 0.6328148
#> [22,] 0.6306604 0.5876178 0.7991622 0.5816543 0.9117592 0.8720326 0.6143732
#> [23,] 0.5637264 0.5463038 0.7613881 0.5602015 0.8824483 0.8270491 0.5339453
#> [24,] 0.5167742 0.4872856 0.7942997 0.5848966 0.9141447 0.8560029 0.4632844
#> [25,] 0.4655962 0.4226715 0.7807170 0.5368128 0.8939086 0.8311497 0.4319062
#> [26,] 0.4884137 0.4494558 0.7856430 0.5766325 0.8996228 0.8434029 0.4028959
#> [27,] 0.5418102 0.4763979 0.6645791 0.4530207 0.8439474 0.8025723 0.4804143
#> [28,] 0.5351362 0.4955365 0.7775333 0.5680265 0.8938647 0.8557358 0.4814475
#> [29,] 0.5763677 0.5088496 0.7874179 0.5642022 0.8873373 0.8512740 0.4643504
#> [30,] 0.5517337 0.4298322 0.6636695 0.5069974 0.8014608 0.7338635 0.4777507
#> [31,] 0.6259922 0.5941461 0.7446493 0.5371043 0.8674400 0.8433477 0.5788715
#> [32,] 0.5191733 0.5173059 0.7340548 0.4754359 0.9044762 0.8475587 0.4748812
#> [33,] 0.4438070 0.4075180 0.7199797 0.5271631 0.8615324 0.7932157 0.3440015
#> [34,] 0.5707098 0.5766652 0.7796474 0.5692689 0.9090154 0.8563669 0.5241530
#> [35,] 0.4586265 0.3827785 0.6554023 0.4780873 0.7984678 0.7306200 0.4135537
#> [36,] 0.6427583 0.5698030 0.8287216 0.5960616 0.9322600 0.8920076 0.5765528
#> [37,] 0.6265366 0.5718263 0.7645923 0.5561237 0.8909811 0.8345318 0.6082882
#> [38,] 0.6598580 0.6638863 0.8219588 0.5730416 0.9274463 0.8877846 0.6357029
#> [39,] 0.5751397 0.5394279 0.6973285 0.4712447 0.8648817 0.8231211 0.5482480
#> [40,] 0.5760599 0.5262485 0.7733109 0.5742300 0.8715315 0.8129504 0.5184744
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7284014 0.6869834 0.8081682
#>  [2,] 0.7231256 0.6665279 0.8163709
#>  [3,] 0.6562217 0.6622847 0.7073454
#>  [4,] 0.6910338 0.7325833 0.8375359
#>  [5,] 0.5545961 0.4865382 0.5511103
#>  [6,] 0.6937132 0.6307758 0.7888270
#>  [7,] 0.6903677 0.5877167 0.7344506
#>  [8,] 0.6841590 0.7324227 0.7868785
#>  [9,] 0.7369035 0.7490626 0.8797773
#> [10,] 0.7306599 0.7101811 0.8504037
#> [11,] 0.7480504 0.6951409 0.8537276
#> [12,] 0.6908536 0.7211690 0.8251874
#> [13,] 0.7243095 0.6632876 0.8146847
#> [14,] 0.6846527 0.7234801 0.8351688
#> [15,] 0.7308294 0.7660937 0.8812465
#> [16,] 0.7088447 0.7293016 0.8464945
#> [17,] 0.5861485 0.4693919 0.7270624
#> [18,] 0.6700735 0.5164408 0.6674732
#> [19,] 0.6840476 0.6666497 0.7795290
#> [20,] 0.6174583 0.4982855 0.6415407
#> [21,] 0.7344079 0.6881685 0.8353556
#> [22,] 0.7401816 0.6585432 0.7815839
#> [23,] 0.6715669 0.6581618 0.7666825
#> [24,] 0.7170324 0.6791045 0.7705448
#> [25,] 0.6541054 0.6791538 0.7739805
#> [26,] 0.6746584 0.6858230 0.7936264
#> [27,] 0.6698568 0.4895151 0.6904694
#> [28,] 0.6870537 0.6436181 0.7919205
#> [29,] 0.6623388 0.7010304 0.7885510
#> [30,] 0.6521642 0.4954449 0.5908501
#> [31,] 0.6823198 0.5991388 0.7636952
#> [32,] 0.6711572 0.6275928 0.8147296
#> [33,] 0.6318457 0.6472937 0.7465271
#> [34,] 0.7105973 0.6909374 0.8396553
#> [35,] 0.6293542 0.4988733 0.6181971
#> [36,] 0.7550946 0.7243087 0.8367326
#> [37,] 0.7277067 0.6201423 0.7674012
#> [38,] 0.7298347 0.7408387 0.8749342
#> [39,] 0.6900270 0.5289291 0.7415035
#> [40,] 0.6979050 0.6815113 0.7774620
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.3042870 0.6957130
#>  [2,] 0.3071486 0.6928514
#>  [3,] 0.3293317 0.6706683
#>  [4,] 0.3120364 0.6879636
#>  [5,] 0.3575838 0.6424162
#>  [6,] 0.3162517 0.6837483
#>  [7,] 0.3266811 0.6733189
#>  [8,] 0.3125354 0.6874646
#>  [9,] 0.3001418 0.6998582
#> [10,] 0.3029655 0.6970345
#> [11,] 0.3007684 0.6992316
#> [12,] 0.3112350 0.6887650
#> [13,] 0.3068982 0.6931018
#> [14,] 0.3128875 0.6871125
#> [15,] 0.3020832 0.6979168
#> [16,] 0.3076962 0.6923038
#> [17,] 0.3491584 0.6508416
#> [18,] 0.3366711 0.6633289
#> [19,] 0.3177825 0.6822175
#> [20,] 0.3457689 0.6542311
#> [21,] 0.3043523 0.6956477
#> [22,] 0.3077765 0.6922235
#> [23,] 0.3177957 0.6822043
#> [24,] 0.3169071 0.6830929
#> [25,] 0.3254091 0.6745909
#> [26,] 0.3207808 0.6792192
#> [27,] 0.3374181 0.6625819
#> [28,] 0.3187018 0.6812982
#> [29,] 0.3151795 0.6848205
#> [30,] 0.3400441 0.6599559
#> [31,] 0.3170171 0.6829829
#> [32,] 0.3246672 0.6753328
#> [33,] 0.3335673 0.6664327
#> [34,] 0.3113713 0.6886287
#> [35,] 0.3479471 0.6520529
#> [36,] 0.3016069 0.6983931
#> [37,] 0.3132848 0.6867152
#> [38,] 0.2992488 0.7007512
#> [39,] 0.3282437 0.6717563
#> [40,] 0.3155491 0.6844509
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>              [,1]         [,2]      [,3]        [,4]     [,5]      [,6]
#>  [1,]  0.70764986  0.437209287 1.3970225  0.29179366 2.337620 1.9943321
#>  [2,]  0.66129217  0.603991237 1.1962536  0.31719060 1.869885 1.4808950
#>  [3,] -0.14161757 -0.394707556 1.1959464  0.25804270 1.779890 1.3818457
#>  [4,]  0.16893255  0.059947379 1.4884107  0.25935874 2.502294 1.9884543
#>  [5,] -0.28597582 -0.717577696 0.3926199 -0.04631458 1.206386 0.7771415
#>  [6,]  0.43946335  0.328337033 1.1513092  0.08147201 2.016097 1.7093867
#>  [7,]  0.06501142  0.003157256 1.0075414  0.12316504 1.969531 1.4521044
#>  [8,]  0.20896719  0.045700565 1.3758069  0.43973039 2.157393 1.7193024
#>  [9,]  0.56138280  0.531352740 1.6421905  0.27352481 2.726894 2.2508641
#> [10,]  0.64745874  0.577420419 1.4236615  0.25182650 2.334137 1.9283474
#> [11,]  0.78956354  0.737227041 1.3963495  0.21606583 2.312531 2.0202812
#> [12,]  0.26093641  0.136802725 1.4485763  0.32330030 2.213376 1.7755781
#> [13,]  0.71619568  0.607815294 1.1717687  0.30462962 1.843807 1.5338363
#> [14,]  0.16837400  0.154454719 1.4353872  0.23723291 2.409901 2.0774992
#> [15,]  0.40221131  0.395008562 1.6893024  0.29814524 2.757173 2.2634501
#> [16,]  0.27487452  0.217243424 1.5928632  0.29455841 2.589735 2.0230810
#> [17,] -0.22938312 -0.050642539 0.5122608 -0.40895891 1.613481 1.2245065
#> [18,]  0.06604795 -0.112512995 0.7397557 -0.05997454 1.775430 1.3682500
#> [19,]  0.17437585 -0.070569294 1.3153431  0.22853813 2.238182 1.7511103
#> [20,] -0.15788460 -0.381133913 0.7209661 -0.04945789 1.436349 0.9833985
#> [21,]  0.66530048  0.566582916 1.3725588  0.23038198 2.292345 1.8960254
#> [22,]  0.53505113  0.354126006 1.3810660  0.32956797 2.335307 1.9190510
#> [23,]  0.25629950  0.185747556 1.1603050  0.24198003 2.015822 1.5648563
#> [24,]  0.06712201 -0.050868532 1.3510406  0.34290732 2.365326 1.7824803
#> [25,] -0.13783299 -0.311816238 1.2698496  0.14751820 2.131302 1.5937971
#> [26,] -0.04635359 -0.202869665 1.2988597  0.30896440 2.193040 1.6837685
#> [27,]  0.16763243 -0.094478532 0.6837678 -0.18847316 1.687897 1.4024493
#> [28,]  0.14077687 -0.017854513 1.2513492  0.27380363 2.130840 1.7803149
#> [29,]  0.30787988  0.035402237 1.3094309  0.25823452 2.063826 1.7446285
#> [30,]  0.20767817 -0.282535606 0.6796901  0.02799139 1.395450 1.0143140
#> [31,]  0.51506117  0.381131893 1.0702753  0.14869067 1.878511 1.6833505
#> [32,]  0.07673065  0.069251199 1.0152933 -0.09833555 2.247981 1.7155805
#> [33,] -0.22572548 -0.374235372 0.9443610  0.10875957 1.828077 1.3444192
#> [34,]  0.28474786  0.309098307 1.2636126  0.27886887 2.301672 1.7854365
#> [35,] -0.16587329 -0.477771492 0.6428718 -0.08770716 1.376746 0.9977705
#> [36,]  0.58735633  0.281047589 1.5765943  0.38908155 2.621934 2.1114135
#> [37,]  0.51738733  0.289306367 1.1780241  0.22544497 2.100802 1.6180917
#> [38,]  0.66266138  0.680661334 1.5296752  0.29427163 2.548108 2.0683093
#> [39,]  0.30285244  0.158039644 0.8346086 -0.11514806 1.856442 1.5376382
#> [40,]  0.30661959  0.105090737 1.2271015  0.29913071 1.914569 1.4692964
#>              [,7]      [,8]         [,9]     [,10]
#>  [1,]  0.44754114 0.9865271  0.786054132 1.4381512
#>  [2,]  0.67071911 0.9600188  0.692522677 1.4919508
#>  [3,] -0.49647383 0.6465015  0.673492402 0.8825259
#>  [4,] -0.20392909 0.8049569  1.007768927 1.6400072
#>  [5,] -0.69470374 0.2192588 -0.053860073 0.2051579
#>  [6,]  0.24010308 0.8175368  0.535546482 1.3178693
#>  [7,]  0.03012508 0.8018390  0.354534323 1.0173217
#>  [8,] -0.15645907 0.7729514  1.006949482 1.3062112
#>  [9,]  0.36568720 1.0299362  1.093619033 1.9903233
#> [10,]  0.51376798 0.9979734  0.896263912 1.7377708
#> [11,]  0.76987781 1.0882413  0.824264851 1.7641414
#> [12,] -0.01302970 0.8041129  0.950267424 1.5518956
#> [13,]  0.64832866 0.9659400  0.677979590 1.4807424
#> [14,] -0.24138210 0.7752370  0.961790046 1.6227118
#> [15,]  0.15431827 0.9988347  1.186383944 2.0042878
#> [16,]  0.17440022 0.8897797  0.991082199 1.7073673
#> [17,] -0.28930868 0.3480659 -0.122585733 0.9797692
#> [18,] -0.16212558 0.7085174  0.065787105 0.6967788
#> [19,] -0.02173885 0.7724359  0.693070811 1.2629238
#> [20,] -0.24418167 0.4787739 -0.006858043 0.5820574
#> [21,]  0.54431158 1.0171029  0.791570614 1.6240694
#> [22,]  0.46573241 1.0469125  0.656808826 1.2749204
#> [23,]  0.13599034 0.7152806  0.655113111 1.1896727
#> [24,] -0.14712706 0.9297888  0.749659409 1.2113898
#> [25,] -0.27407831 0.6371343  0.749885853 1.2309252
#> [26,] -0.39341323 0.7293308  0.780662916 1.3469248
#> [27,] -0.07838303 0.7075374 -0.041945791 0.8023146
#> [28,] -0.07424427 0.7863808  0.591102487 1.3365411
#> [29,] -0.14284085 0.6737341  0.852209334 1.3162130
#> [30,] -0.08905596 0.6285660 -0.018220987 0.3674806
#> [31,]  0.31814257 0.7644533  0.401878223 1.1730460
#> [32,] -0.10056003 0.7134234  0.521903676 1.4810400
#> [33,] -0.64551267 0.5401431  0.607164641 1.0801752
#> [34,]  0.09668741 0.8982867  0.804505456 1.6556654
#> [35,] -0.34929381 0.5294475 -0.004506734 0.4819029
#> [36,]  0.30863807 1.1259709  0.965936072 1.6341149
#> [37,]  0.44012236 0.9830182  0.490152173 1.1936947
#> [38,]  0.55676158 0.9937839  1.050332025 1.9453088
#> [39,]  0.19359420 0.8002456  0.115845674 1.0537976
#> [40,]  0.07393112 0.8373414  0.760725862 1.2509371
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.3624104 0.4645555
#>  [2,] -0.3622198 0.4512641
#>  [3,] -0.2991016 0.4121078
#>  [4,] -0.3347333 0.4558829
#>  [5,] -0.2567701 0.3290965
#>  [6,] -0.3356855 0.4353662
#>  [7,] -0.3091423 0.4140921
#>  [8,] -0.3396537 0.4486388
#>  [9,] -0.3646070 0.4820157
#> [10,] -0.3640427 0.4691734
#> [11,] -0.3702369 0.4734048
#> [12,] -0.3408988 0.4534531
#> [13,] -0.3636308 0.4510298
#> [14,] -0.3341899 0.4524644
#> [15,] -0.3570206 0.4803769
#> [16,] -0.3443349 0.4665768
#> [17,] -0.2663639 0.3563767
#> [18,] -0.2919819 0.3861824
#> [19,] -0.3247216 0.4392597
#> [20,] -0.2738761 0.3638138
#> [21,] -0.3617036 0.4649538
#> [22,] -0.3517889 0.4587463
#> [23,] -0.3317361 0.4321842
#> [24,] -0.3241855 0.4438368
#> [25,] -0.3028677 0.4261556
#> [26,] -0.3164028 0.4337832
#> [27,] -0.2915982 0.3832233
#> [28,] -0.3244252 0.4353192
#> [29,] -0.3355985 0.4404162
#> [30,] -0.2926961 0.3704018
#> [31,] -0.3389898 0.4285245
#> [32,] -0.3099132 0.4224919
#> [33,] -0.2920399 0.4000547
#> [34,] -0.3427282 0.4509878
#> [35,] -0.2686026 0.3594727
#> [36,] -0.3623804 0.4772773
#> [37,] -0.3430172 0.4417900
#> [38,] -0.3714365 0.4794409
#> [39,] -0.3105965 0.4055430
#> [40,] -0.3361675 0.4381349
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,] -0.3042870  0.3042870
#>  [2,] -0.3071486  0.3071486
#>  [3,] -0.3293317  0.3293317
#>  [4,]  0.6879636 -0.6879636
#>  [5,]  0.6424162 -0.6424162
#>  [6,]  0.6837483 -0.6837483
#>  [7,] -0.3266811  0.3266811
#>  [8,]  0.6874646 -0.6874646
#>  [9,]  0.6998582 -0.6998582
#> [10,] -0.3029655  0.3029655
#> [11,]  0.6992316 -0.6992316
#> [12,] -0.3112350  0.3112350
#> [13,] -0.3068982  0.3068982
#> [14,] -0.3128875  0.3128875
#> [15,]  0.6979168 -0.6979168
#> [16,]  0.6923038 -0.6923038
#> [17,]  0.6508416 -0.6508416
#> [18,] -0.3366711  0.3366711
#> [19,] -0.3177825  0.3177825
#> [20,] -0.3457689  0.3457689
#> [21,]  0.6956477 -0.6956477
#> [22,] -0.3077765  0.3077765
#> [23,]  0.6822043 -0.6822043
#> [24,] -0.3169071  0.3169071
#> [25,]  0.6745909 -0.6745909
#> [26,]  0.6792192 -0.6792192
#> [27,]  0.6625819 -0.6625819
#> [28,] -0.3187018  0.3187018
#> [29,] -0.3151795  0.3151795
#> [30,] -0.3400441  0.3400441
#> [31,] -0.3170171  0.3170171
#> [32,]  0.6753328 -0.6753328
#> [33,] -0.3335673  0.3335673
#> [34,]  0.6886287 -0.6886287
#> [35,]  0.6520529 -0.6520529
#> [36,]  0.6983931 -0.6983931
#> [37,] -0.3132848  0.3132848
#> [38,]  0.7007512 -0.7007512
#> [39,]  0.6717563 -0.6717563
#> [40,] -0.3155491  0.3155491
#> 
#> $L
#> [1] 0.6954504 0.6908370 0.6928479 0.6961653 0.7942166 0.7643806
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3913043