Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10          V11         V12         V13         V14
#>  [1,] -0.3225942 -0.133643036  0.012541122  0.04552239 -0.10791995 -0.02667513
#>  [2,] -0.2392850 -0.121170093 -0.189856147 -0.14541427 -0.13184407 -0.19678231
#>  [3,] -0.3205834 -0.080290943 -0.011948246  0.04549470  0.01948906  0.08387862
#>  [4,] -0.3862229  0.007915585 -0.139352670 -0.08412676 -0.04949637 -0.10066488
#>  [5,] -0.3064605 -0.017524657  0.003356505 -0.08524649 -0.14194014 -0.02276354
#>  [6,] -0.2601608 -0.141991042 -0.158398958 -0.18648129 -0.16316281 -0.08151469
#>  [7,] -0.3309143 -0.104131148 -0.068394995 -0.20437914 -0.09583534 -0.19974210
#>  [8,] -0.2930577 -0.239668380 -0.198206618 -0.19232028 -0.13959472 -0.08003396
#>  [9,] -0.2995372 -0.176683979 -0.117745223 -0.15387638 -0.05998694 -0.13848567
#> [10,] -0.2645107 -0.194211467 -0.092792487 -0.13064269 -0.12243812 -0.13170203
#>                V15          V16         V17           V18          V19
#>  [1,] -0.030374216  0.017908621  0.00966211  0.0973791973  0.112598539
#>  [2,] -0.118747022 -0.116925208 -0.03683739 -0.0976880478 -0.153445061
#>  [3,] -0.051264284  0.146365436  0.04925068  0.0870870971  0.199419877
#>  [4,]  0.032543576 -0.070969777 -0.06239928  0.0009053896 -0.008545181
#>  [5,] -0.006705336  0.002733869  0.08566595  0.0981697850  0.124976794
#>  [6,] -0.106824552 -0.037908712 -0.04494702  0.0445460083  0.058047374
#>  [7,] -0.054036053  0.015506096 -0.02763586 -0.1258883405 -0.068878006
#>  [8,] -0.109555444 -0.037459764 -0.06376055 -0.1362102962 -0.018331552
#>  [9,] -0.072243619 -0.130715290 -0.13167961 -0.0989383309 -0.106678332
#> [10,] -0.066839110 -0.131904073  0.01222979 -0.1009669128 -0.028498328
#>               V2          V20          V21         V22          V23         V24
#>  [1,] -0.3333955  0.063199140  0.135104700  0.15557898  0.252169238  0.28842611
#>  [2,] -0.2417575  0.031753761  0.088688464  0.06928879  0.076122318 -0.04364943
#>  [3,] -0.3361740  0.102405715  0.062490652  0.15791841  0.128040429  0.31735642
#>  [4,] -0.2763183  0.016430374  0.167176374  0.20265083  0.099595797  0.06480542
#>  [5,] -0.1885802  0.155218646  0.087507849  0.14674328  0.120738325  0.11538262
#>  [6,] -0.3379563  0.075939654  0.150886731 -0.01731564  0.005150342  0.12650360
#>  [7,] -0.3222387 -0.079321618 -0.051253098  0.07716360 -0.035584782  0.09400441
#>  [8,] -0.2061581 -0.060675683  0.054571408  0.06374603 -0.012561755  0.04209493
#>  [9,] -0.2536110 -0.005099207  0.008097964  0.03871504  0.155749822  0.03990368
#> [10,] -0.3588563 -0.089997353  0.049192941  0.03620414  0.020326468  0.01288890
#>                V25        V26          V27         V28         V29         V3
#>  [1,]  0.301957955 0.24203234  0.164196619  0.11746342  0.12995777 -0.2590071
#>  [2,]  0.090100733 0.01222982 -0.019072345  0.12710402 -0.02045613 -0.2985877
#>  [3,]  0.307177446 0.16062079  0.122659334  0.19667747  0.10965416 -0.2638697
#>  [4,]  0.059947122 0.16806248  0.198310259  0.10502395  0.12599631 -0.3109179
#>  [5,]  0.230075849 0.21716604  0.200655777  0.11458081  0.09746327 -0.2732149
#>  [6,]  0.085952828 0.19850947  0.159693513  0.14027032  0.19232729 -0.2485921
#>  [7,]  0.062321050 0.10345427  0.103475214 -0.05112142 -0.01110416 -0.1821189
#>  [8,]  0.096653348 0.08699721  0.133580702  0.07262435 -0.04768690 -0.3593422
#>  [9,] -0.006646528 0.16196231 -0.001801012  0.14565479  0.09399779 -0.2878982
#> [10,]  0.030590473 0.09832741  0.159246117  0.18413049  0.12650328 -0.1732790
#>               V30         V31          V32          V33         V34
#>  [1,]  0.07236864  0.03575219 -0.025731060 -0.007251394 -0.03609364
#>  [2,] -0.14394026 -0.15859687 -0.154359467 -0.109308628 -0.08204274
#>  [3,]  0.17218633  0.04831178  0.074540224  0.127266334  0.02949913
#>  [4,]  0.09120283  0.14551476  0.017831545 -0.002282996  0.14011663
#>  [5,]  0.17093284  0.14013989  0.091228458  0.008406137 -0.04373182
#>  [6,]  0.17301490 -0.02430370  0.121708564  0.098573389  0.04733140
#>  [7,] -0.10748639 -0.01620566  0.073185714 -0.070324885 -0.14563732
#>  [8,] -0.07292912 -0.02133562 -0.119245638 -0.058108982 -0.20014174
#>  [9,]  0.02479310 -0.08601757  0.022946311  0.001295812 -0.05224669
#> [10,]  0.05635581  0.06198715  0.003820913  0.042550253  0.06656085
#>                V35          V36          V37         V38          V39
#>  [1,] -0.047153439  0.081274743  0.023879517  0.08540647 -0.036263276
#>  [2,]  0.002589782 -0.133635727 -0.137760637 -0.03215814 -0.033903439
#>  [3,]  0.057006491 -0.028373986 -0.002355289  0.04192055  0.069241982
#>  [4,]  0.073305182  0.141873207  0.084054068 -0.07856264 -0.080483872
#>  [5,]  0.032524508 -0.019582383 -0.055747212 -0.09184929  0.031169727
#>  [6,] -0.007523910  0.002260674  0.044264793  0.07400168 -0.002349768
#>  [7,] -0.083652263 -0.208985738 -0.129197990 -0.09231774 -0.017129978
#>  [8,] -0.117267802 -0.078996250 -0.111687871 -0.17504597 -0.075932292
#>  [9,] -0.074890725  0.010195342  0.012243056 -0.12370852 -0.164543445
#> [10,] -0.070562955  0.045690445 -0.054141616 -0.02121126 -0.099638356
#>               V4          V40          V41         V42         V43          V44
#>  [1,] -0.2237072  0.059777931 -0.068922823 -0.04694098  0.01151571 -0.097429094
#>  [2,] -0.3303770 -0.157571579 -0.022405283 -0.18252952 -0.12832851 -0.058082646
#>  [3,] -0.3282960  0.072740913  0.007314968 -0.02199066  0.01178960 -0.031573921
#>  [4,] -0.1540235 -0.041893952 -0.028243702 -0.05178989 -0.08893267 -0.169795676
#>  [5,] -0.1986949 -0.154984162 -0.101829051 -0.07772114 -0.09476216 -0.007277614
#>  [6,] -0.3190353 -0.102103654  0.025450093 -0.03793761 -0.02003387 -0.178945325
#>  [7,] -0.2887053 -0.003004528 -0.125501836 -0.04239604 -0.05918188  0.002306221
#>  [8,] -0.2319480 -0.207019565 -0.023065072 -0.18086073 -0.23842347 -0.211932249
#>  [9,] -0.2674192  0.004742685 -0.023139049 -0.04822234 -0.15100280 -0.177154493
#> [10,] -0.2275575 -0.063115554 -0.106857570  0.03499177 -0.16286974 -0.167519424
#>                V45        V46         V47        V48        V49         V5
#>  [1,] -0.126936615 -0.2059782 -0.07364067 -0.2540661 -0.2331032 -0.1879702
#>  [2,] -0.212766749 -0.1083319 -0.17605472 -0.2181124 -0.2653018 -0.1741450
#>  [3,] -0.013627936 -0.1175751 -0.20737773 -0.1330477 -0.3302928 -0.2475559
#>  [4,] -0.161920167 -0.1580469 -0.07992473 -0.1304033 -0.2440033 -0.2525899
#>  [5,] -0.073063092 -0.2010533 -0.15206242 -0.2800195 -0.2222176 -0.2773911
#>  [6,]  0.005264330 -0.1332648 -0.20387648 -0.1543615 -0.3036535 -0.2423815
#>  [7,] -0.159191498 -0.1328305 -0.20970348 -0.1458174 -0.2745605 -0.1753811
#>  [8,] -0.197087438 -0.1614308 -0.27645264 -0.2571097 -0.2444462 -0.2705575
#>  [9,] -0.175303137 -0.1023228 -0.24586727 -0.2541208 -0.2992939 -0.2170835
#> [10,]  0.007778082 -0.0483684 -0.20107909 -0.3026224 -0.2138128 -0.2870832
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2941422 -0.2876409 -0.3496134 -0.3226279 -0.3931099 -0.3867929
#>  [2,] -0.3620567 -0.2507691 -0.3287794 -0.3849261 -0.3263825 -0.2963805
#>  [3,] -0.3230784 -0.2810460 -0.2880956 -0.4191816 -0.2541792 -0.3663090
#>  [4,] -0.3070694 -0.3277924 -0.3739066 -0.3622144 -0.2655793 -0.3791523
#>  [5,] -0.3864606 -0.3660033 -0.4110771 -0.2845483 -0.2763124 -0.2978761
#>  [6,] -0.2469796 -0.2576728 -0.2312289 -0.3837796 -0.3690385 -0.3686328
#>  [7,] -0.3328461 -0.4086462 -0.4075997 -0.3526529 -0.2840809 -0.3138105
#>  [8,] -0.2338077 -0.2351995 -0.2867063 -0.3087731 -0.3467655 -0.2364662
#>  [9,] -0.2146202 -0.2448059 -0.3564820 -0.2739389 -0.3460072 -0.2892175
#> [10,] -0.3112696 -0.4242810 -0.3318476 -0.2622960 -0.3196639 -0.3461801
#>              V56        V57        V58        V59          V6        V60
#>  [1,] -0.3238118 -0.3505961 -0.3731199 -0.2123526 -0.15264372 -0.3029847
#>  [2,] -0.3119751 -0.2841818 -0.2918611 -0.3864577 -0.17027087 -0.3761932
#>  [3,] -0.3480444 -0.4042523 -0.2536736 -0.2753961 -0.21723338 -0.3914428
#>  [4,] -0.2265818 -0.2752296 -0.3206792 -0.4353552 -0.23523531 -0.3193155
#>  [5,] -0.3127237 -0.2458221 -0.2866419 -0.3397312 -0.26010199 -0.3241506
#>  [6,] -0.3665589 -0.3903614 -0.4234825 -0.4035462 -0.17680065 -0.3149413
#>  [7,] -0.4210976 -0.4161695 -0.3024333 -0.2974545 -0.13605391 -0.3579606
#>  [8,] -0.3536196 -0.2733738 -0.3903144 -0.4263331 -0.21893465 -0.2784680
#>  [9,] -0.3510224 -0.3609171 -0.2424004 -0.2700300 -0.09351791 -0.2904422
#> [10,] -0.2598639 -0.3005672 -0.4336716 -0.2865236 -0.17285087 -0.2957693
#>                V7          V8          V9
#>  [1,] -0.24316914 -0.16420644 -0.04712026
#>  [2,] -0.18595873 -0.10546029 -0.14376049
#>  [3,] -0.23834599 -0.15772679 -0.11661241
#>  [4,] -0.09187153 -0.23141622 -0.02737448
#>  [5,] -0.08841441 -0.07067754 -0.15051838
#>  [6,] -0.14421931 -0.26022669 -0.14877355
#>  [7,] -0.31976882 -0.13207806 -0.17532864
#>  [8,] -0.15570655 -0.24239268 -0.18798819
#>  [9,] -0.21454559 -0.15252812 -0.21563206
#> [10,] -0.19145599 -0.11950967 -0.21197626
#> 
#> $W[[2]]
#>            [,1]       [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.09924737 -0.1202122  0.05575727 -0.07484751  0.07418547  0.004042827
#> [2,] 0.02651662  0.0362168 -0.04621887 -0.03361149 -0.03099885 -0.060432764
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] 0.014754413 -0.09717491 -0.04429999 0.005555948
#> [2,] 0.000800913  0.04538383  0.09742870 0.088154900
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1          V10          V11          V12          V13
#>  [1,]  3.980922e-05 3.412605e-04 3.870817e-04 0.0004817682 0.0005648198
#>  [2,]  2.534500e-04 1.986383e-03 2.522417e-03 0.0025094002 0.0023535736
#>  [3,]  1.291639e-05 1.775751e-04 1.793000e-04 0.0002603717 0.0003455961
#>  [4,]  1.802577e-04 1.282024e-03 1.545813e-03 0.0016164647 0.0016598422
#>  [5,] -7.660674e-06 6.328865e-05 1.628008e-05 0.0001342131 0.0002632157
#>  [6,]  4.020436e-06 1.741992e-04 1.212193e-04 0.0002314640 0.0003938852
#>  [7,]  2.403810e-05 2.178700e-04 1.958920e-04 0.0002376833 0.0003537419
#>  [8,]  2.173685e-04 1.690887e-03 2.143243e-03 0.0021600770 0.0020554168
#>  [9,]  2.697857e-04 2.100828e-03 2.610936e-03 0.0026179707 0.0025235374
#> [10,]  2.249249e-04 1.697867e-03 2.059803e-03 0.0021109991 0.0021191167
#>                V14          V15          V16          V17          V18
#>  [1,] 0.0006525286 0.0007123261 0.0009675787 0.0011319685 0.0012250480
#>  [2,] 0.0018180221 0.0016124857 0.0012736680 0.0012324993 0.0016296730
#>  [3,] 0.0004665333 0.0005302255 0.0007803095 0.0009308921 0.0009943167
#>  [4,] 0.0015445596 0.0015280714 0.0016610637 0.0018030040 0.0020385856
#>  [5,] 0.0004849038 0.0006127292 0.0009998613 0.0012132567 0.0012685032
#>  [6,] 0.0006722093 0.0008529764 0.0013065784 0.0015432378 0.0015788924
#>  [7,] 0.0005200644 0.0006141337 0.0008649730 0.0009891192 0.0010367567
#>  [8,] 0.0016322676 0.0014748011 0.0012386944 0.0012436779 0.0016042857
#>  [9,] 0.0020554698 0.0018560623 0.0016370885 0.0016434291 0.0020385248
#> [10,] 0.0018947681 0.0018268014 0.0018779848 0.0019952627 0.0023168571
#>                V19            V2          V20          V21          V22
#>  [1,] 0.0012880503  6.430249e-05 0.0012139383 0.0011467301 0.0012839021
#>  [2,] 0.0028979415  3.701807e-04 0.0033626042 0.0035893416 0.0036223809
#>  [3,] 0.0009734599  2.419563e-05 0.0008785919 0.0008049736 0.0009449113
#>  [4,] 0.0026446373  2.669615e-04 0.0028315633 0.0029335056 0.0030439959
#>  [5,] 0.0011152210 -2.290101e-06 0.0009140122 0.0007612415 0.0009538448
#>  [6,] 0.0013993093  1.754674e-05 0.0011808168 0.0010435089 0.0012644600
#>  [7,] 0.0010178037  3.573061e-05 0.0009593804 0.0009216006 0.0010464415
#>  [8,] 0.0026794017  3.164632e-04 0.0030328098 0.0032007101 0.0032831655
#>  [9,] 0.0033088151  3.920494e-04 0.0037822799 0.0040406976 0.0041560305
#> [10,] 0.0032120936  3.273234e-04 0.0035231959 0.0036989442 0.0038380759
#>               V23         V24         V25         V26         V27         V28
#>  [1,] 0.001489898 0.001524063 0.001502975 0.001522475 0.001549819 0.001512368
#>  [2,] 0.003438503 0.003287225 0.002518165 0.002565794 0.002744333 0.002750154
#>  [3,] 0.001151880 0.001176027 0.001203979 0.001238503 0.001248060 0.001194578
#>  [4,] 0.003155802 0.003181331 0.002859090 0.002892178 0.002990765 0.002965000
#>  [5,] 0.001258906 0.001302077 0.001398903 0.001414789 0.001399862 0.001321566
#>  [6,] 0.001588280 0.001630636 0.001735941 0.001739836 0.001703148 0.001595755
#>  [7,] 0.001245120 0.001296536 0.001351092 0.001394202 0.001366762 0.001312455
#>  [8,] 0.003196813 0.003077615 0.002436284 0.002477589 0.002635221 0.002634567
#>  [9,] 0.004094687 0.004007401 0.003293651 0.003363063 0.003548775 0.003561700
#> [10,] 0.003907789 0.003897243 0.003415753 0.003468676 0.003590274 0.003562550
#>               V29           V3          V30          V31          V32
#>  [1,] 0.001424711 1.008094e-04 0.0012226289 0.0013402596 0.0011865109
#>  [2,] 0.002626869 4.357613e-04 0.0017562819 0.0009673307 0.0008845768
#>  [3,] 0.001108934 4.984865e-05 0.0009796162 0.0011308544 0.0009880992
#>  [4,] 0.002787827 3.430115e-04 0.0021995095 0.0019603669 0.0017719953
#>  [5,] 0.001230399 2.359798e-05 0.0011416882 0.0014469102 0.0012778461
#>  [6,] 0.001463945 5.165372e-05 0.0013558703 0.0017021142 0.0014967530
#>  [7,] 0.001235792 5.598155e-05 0.0011071421 0.0012359276 0.0010940330
#>  [8,] 0.002519542 3.793863e-04 0.0017269299 0.0011110778 0.0010095228
#>  [9,] 0.003374294 4.756223e-04 0.0023916462 0.0016899003 0.0015450683
#> [10,] 0.003358338 4.104311e-04 0.0025722362 0.0021601296 0.0019468946
#>                V33          V34           V35           V36           V37
#>  [1,] 0.0010599431 0.0009896969  9.987990e-04  1.082260e-03  1.029879e-03
#>  [2,] 0.0007102682 0.0002079658 -4.007396e-04 -5.896389e-04 -5.301181e-04
#>  [3,] 0.0008953327 0.0008688655  9.178833e-04  1.014283e-03  9.594918e-04
#>  [4,] 0.0015730414 0.0012727408  1.004979e-03  9.874807e-04  9.167714e-04
#>  [5,] 0.0011598783 0.0011736142  1.312027e-03  1.466108e-03  1.401860e-03
#>  [6,] 0.0013299092 0.0013256342  1.478597e-03  1.644866e-03  1.550792e-03
#>  [7,] 0.0010064480 0.0009765222  1.023146e-03  1.065164e-03  9.650811e-04
#>  [8,] 0.0008225927 0.0003734602 -1.433472e-04 -2.892000e-04 -2.617970e-04
#>  [9,] 0.0012863425 0.0007050673  7.408552e-05 -6.338674e-05 -4.675019e-05
#> [10,] 0.0017055376 0.0012917799  8.808985e-04  8.188045e-04  7.531907e-04
#>                V38          V39           V4          V40          V41
#>  [1,] 0.0007868553 0.0007645530 1.136301e-04 0.0007981264 0.0005884202
#>  [2,] 0.0003214796 0.0016114536 6.346905e-04 0.0012809248 0.0011594829
#>  [3,] 0.0006801669 0.0005780284 4.726009e-05 0.0006426361 0.0004608312
#>  [4,] 0.0010496726 0.0015707811 4.566911e-04 0.0014381784 0.0012005574
#>  [5,] 0.0009189047 0.0006442491 7.283093e-07 0.0007743575 0.0005292286
#>  [6,] 0.0009955200 0.0007311650 4.185382e-05 0.0008924226 0.0006130321
#>  [7,] 0.0006744695 0.0005703571 6.490241e-05 0.0006241753 0.0004392566
#>  [8,] 0.0003799609 0.0014583348 5.539643e-04 0.0012076361 0.0010657630
#>  [9,] 0.0006773112 0.0019484786 6.762851e-04 0.0016182918 0.0014196965
#> [10,] 0.0010710563 0.0018948731 5.620464e-04 0.0016789726 0.0014061258
#>                V42          V43          V44           V45           V46
#>  [1,] 0.0004325156 3.412281e-04 0.0003477745  2.040800e-04  1.381281e-04
#>  [2,] 0.0016512281 1.613655e-03 0.0015287657  1.710505e-03  1.155226e-03
#>  [3,] 0.0002747304 1.869343e-04 0.0001957033  4.813518e-05  3.275287e-05
#>  [4,] 0.0012855650 1.163055e-03 0.0011149086  1.098352e-03  7.507599e-04
#>  [5,] 0.0002092514 7.278025e-05 0.0001048174 -1.300625e-04 -8.449477e-05
#>  [6,] 0.0002406511 1.073092e-04 0.0001665360 -7.113845e-05 -4.413579e-05
#>  [7,] 0.0002568450 1.789635e-04 0.0002002242  8.773173e-05  6.739961e-05
#>  [8,] 0.0014213894 1.364232e-03 0.0013070492  1.425624e-03  9.670297e-04
#>  [9,] 0.0018519568 1.790438e-03 0.0017229389  1.853432e-03  1.238398e-03
#> [10,] 0.0016017816 1.492232e-03 0.0014365969  1.443025e-03  9.674357e-04
#>                V47           V48           V49           V5          V50
#>  [1,] 1.734477e-04  1.156002e-04  6.511644e-05 1.751409e-04 3.541184e-05
#>  [2,] 1.020405e-03  8.946393e-04  5.408007e-04 7.501058e-04 1.283406e-04
#>  [3,] 7.620083e-05  3.429161e-05  1.909732e-05 1.001015e-04 2.269951e-05
#>  [4,] 6.938777e-04  5.748328e-04  3.350252e-04 5.550266e-04 1.031391e-04
#>  [5,] 1.998695e-09 -4.963384e-05 -3.109856e-05 6.685982e-05 1.844863e-05
#>  [6,] 4.602420e-05 -1.837044e-05 -1.315839e-05 1.171678e-04 2.821456e-05
#>  [7,] 9.345697e-05  4.991780e-05  2.763566e-05 1.126706e-04 2.893242e-05
#>  [8,] 8.756466e-04  7.531621e-04  4.552156e-04 6.699759e-04 1.141302e-04
#>  [9,] 1.102518e-03  9.584265e-04  5.735714e-04 8.114952e-04 1.454399e-04
#> [10,] 8.868248e-04  7.516418e-04  4.449526e-04 6.863488e-04 1.288202e-04
#>                 V51          V52          V53          V54          V55
#>  [1,]  1.990981e-05 2.521458e-05 2.238532e-05 1.896867e-05 1.838947e-05
#>  [2,]  1.583666e-04 1.494253e-04 6.509883e-05 7.182830e-05 3.141973e-05
#>  [3,]  6.725419e-06 1.138567e-05 1.530116e-05 1.205736e-05 1.403385e-05
#>  [4,]  9.937582e-05 1.033956e-04 5.843998e-05 5.698771e-05 3.859635e-05
#>  [5,] -5.828840e-06 1.045755e-06 1.425281e-05 9.216921e-06 1.523420e-05
#>  [6,] -2.451974e-06 8.168846e-06 1.788085e-05 1.257299e-05 1.844476e-05
#>  [7,]  1.013732e-05 1.676284e-05 1.789279e-05 1.517272e-05 1.514484e-05
#>  [8,]  1.347267e-04 1.307419e-04 5.914223e-05 6.343542e-05 2.905100e-05
#>  [9,]  1.653883e-04 1.605527e-04 7.594405e-05 8.166351e-05 4.121156e-05
#> [10,]  1.296259e-04 1.313944e-04 7.093704e-05 7.249968e-05 4.375555e-05
#>                V56          V57          V58          V59           V6
#>  [1,] 1.162412e-05 2.005690e-05 1.440803e-05 1.654796e-05 0.0002651157
#>  [2,] 2.730655e-05 2.533222e-05 5.352254e-05 5.234765e-05 0.0006294332
#>  [3,] 8.305562e-06 1.617122e-05 9.125152e-06 1.051716e-05 0.0001965275
#>  [4,] 2.919521e-05 3.580943e-05 4.169931e-05 4.544938e-05 0.0005734134
#>  [5,] 8.396352e-06 1.988778e-05 7.203766e-06 9.463322e-06 0.0002196429
#>  [6,] 1.174487e-05 2.549024e-05 8.864988e-06 1.270111e-05 0.0002876363
#>  [7,] 1.062471e-05 1.801308e-05 9.689467e-06 1.258265e-05 0.0001992601
#>  [8,] 2.447653e-05 2.575788e-05 4.674702e-05 4.703408e-05 0.0005897637
#>  [9,] 3.365554e-05 3.568367e-05 5.847808e-05 5.933180e-05 0.0007319379
#> [10,] 3.432992e-05 4.038767e-05 5.165799e-05 5.494512e-05 0.0006893548
#>                V60           V7           V8           V9
#>  [1,] 1.053940e-05 0.0002872335 0.0002362539 0.0002901214
#>  [2,] 2.420223e-05 0.0007015383 0.0008716871 0.0013769945
#>  [3,] 7.052468e-06 0.0002121622 0.0001590253 0.0001702914
#>  [4,] 2.701637e-05 0.0006294742 0.0006458291 0.0009607191
#>  [5,] 6.876448e-06 0.0002355869 0.0001435713 0.0001048854
#>  [6,] 9.503030e-06 0.0003123899 0.0002108372 0.0001971384
#>  [7,] 8.569070e-06 0.0002245536 0.0001661878 0.0001979465
#>  [8,] 2.178749e-05 0.0006554810 0.0007623419 0.0011816529
#>  [9,] 2.887465e-05 0.0008094679 0.0009542445 0.0014934275
#> [10,] 3.012701e-05 0.0007609493 0.0008218824 0.0012422670
#> 
#> $vW[[2]]
#>             [,1]         [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] -0.04342684 -0.009362769 -0.04592552 -0.0324188 -0.03889182 -0.02579138
#> [2,]  0.04342684  0.009362769  0.04592552  0.0324188  0.03889182  0.02579138
#>              [,7]         [,8]        [,9]       [,10]
#> [1,] -0.009703546 -0.005634419 -0.01054588 -0.01810513
#> [2,]  0.009703546  0.005634419  0.01054588  0.01810513
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6569914 0.3344345 0.5825949 0.6332123 0.5549267 0.4741752 0.1823415
#>  [8] 0.3744935 0.2925890 0.5736653
#> 
#> $B[[2]]
#> [1]  0.10915972 -0.05968121
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.002373056 0.003646599 0.001911380 0.004347955 0.002257910 0.002709911
#>  [7] 0.002061872 0.003538508 0.004773653 0.005085268
#> 
#> $vB[[2]]
#> [1] -0.05588054  0.05588054
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#>  [2,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#>  [3,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#>  [4,] 0.0430 0.7106 0.7342 0.5033 0.3000 0.1951 0.2767 0.3737 0.2507 0.2507
#>  [5,] 0.0414 0.1975 0.2309 0.3025 0.3938 0.5050 0.5872 0.6610 0.7417 0.8006
#>  [6,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#>  [7,] 0.0119 0.0864 0.2143 0.3720 0.2665 0.2113 0.1103 0.1136 0.1934 0.4142
#>  [8,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#>  [9,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [10,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [11,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#> [12,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [13,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [14,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [15,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [16,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [17,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [18,] 0.0125 0.1302 0.0888 0.0500 0.0628 0.1274 0.0801 0.0742 0.2048 0.2950
#> [19,] 0.0072 0.2318 0.2472 0.2880 0.2126 0.0708 0.1194 0.2808 0.4221 0.5279
#> [20,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [21,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [22,] 0.0299 0.0998 0.0523 0.0904 0.2655 0.3099 0.3520 0.3892 0.3962 0.2449
#> [23,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [24,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [25,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [26,] 0.0156 0.1567 0.2476 0.2783 0.2896 0.2956 0.3189 0.1892 0.1730 0.2226
#> [27,] 0.0189 0.2522 0.2607 0.3710 0.3906 0.2672 0.2716 0.4183 0.6988 0.5733
#> [28,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#> [29,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [30,] 0.0293 0.0751 0.0528 0.1209 0.1763 0.2039 0.2727 0.2321 0.2676 0.2934
#> [31,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#> [32,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [33,] 0.0712 0.2127 0.2891 0.3985 0.4576 0.5821 0.5027 0.1930 0.2579 0.3177
#> [34,] 0.0158 0.1923 0.2361 0.2719 0.3049 0.2986 0.2226 0.1745 0.2459 0.3100
#> [35,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [36,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [37,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#> [38,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#> [39,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#> [40,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#>  [2,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#>  [3,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#>  [4,] 0.3292 0.0902 0.4871 0.6527 0.8454 0.9739 1.0000 0.6665 0.5323 0.4024
#>  [5,] 0.8456 0.0436 0.7939 0.8804 0.8384 0.7852 0.8479 0.7434 0.6433 0.5514
#>  [6,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#>  [7,] 0.3279 0.0582 0.6222 0.7468 0.7676 0.7867 0.8253 1.0000 0.9481 0.7539
#>  [8,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#>  [9,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [10,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [11,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#> [12,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [13,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [14,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [15,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [16,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [17,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [18,] 0.3193 0.0152 0.4567 0.5959 0.7101 0.8225 0.8425 0.9065 0.9802 1.0000
#> [19,] 0.5857 0.0027 0.6153 0.6753 0.7873 0.8974 0.9828 1.0000 0.8460 0.6055
#> [20,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [21,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [22,] 0.2355 0.0688 0.3045 0.3112 0.4698 0.5534 0.4532 0.4464 0.4670 0.4621
#> [23,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [24,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [25,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [26,] 0.2427 0.0210 0.3149 0.4102 0.3808 0.4896 0.6292 0.7519 0.7985 0.8830
#> [27,] 0.2226 0.0308 0.2631 0.7473 0.7263 0.3393 0.2824 0.6053 0.5897 0.4967
#> [28,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#> [29,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [30,] 0.3295 0.0378 0.4910 0.5402 0.6257 0.6826 0.7527 0.8504 0.8938 0.9928
#> [31,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#> [32,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [33,] 0.2745 0.0901 0.6186 0.8958 0.7442 0.5188 0.2811 0.1773 0.6607 0.7576
#> [34,] 0.3572 0.0239 0.4283 0.4268 0.3735 0.4585 0.6094 0.7221 0.7595 0.8706
#> [35,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [36,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [37,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#> [38,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#> [39,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#> [40,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#>  [2,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#>  [3,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#>  [4,] 0.3444 0.4239 0.0833 0.4182 0.4393 0.1162 0.4336 0.6553 0.6172 0.4373
#>  [5,] 0.3519 0.3168 0.0447 0.3346 0.2056 0.1032 0.3168 0.4040 0.4282 0.4538
#>  [6,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#>  [7,] 0.6008 0.5437 0.0623 0.5387 0.5619 0.5141 0.6084 0.5621 0.5956 0.6078
#>  [8,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#>  [9,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [10,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [11,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#> [12,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [13,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [14,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [15,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [16,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [17,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [18,] 0.8752 0.7583 0.0218 0.6616 0.5786 0.5128 0.4776 0.4994 0.5197 0.5071
#> [19,] 0.3036 0.0144 0.0089 0.2526 0.4335 0.4918 0.5409 0.5961 0.5248 0.3777
#> [20,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [21,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [22,] 0.6988 0.7626 0.0992 0.7025 0.7382 0.7446 0.7927 0.5227 0.3967 0.3042
#> [23,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [24,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [25,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [26,] 0.9915 0.9223 0.0282 0.6981 0.6167 0.5069 0.3921 0.3524 0.2183 0.1245
#> [27,] 0.8616 0.8339 0.0197 0.4084 0.2268 0.1745 0.0507 0.1588 0.3040 0.1369
#> [28,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#> [29,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [30,] 0.9134 0.7080 0.0257 0.6318 0.6126 0.4638 0.2797 0.1721 0.1665 0.2561
#> [31,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#> [32,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [33,] 0.5122 0.4701 0.1276 0.5479 0.4347 0.1276 0.0846 0.0927 0.0313 0.0998
#> [34,] 1.0000 0.9815 0.0150 0.7187 0.5848 0.4192 0.3756 0.3263 0.1944 0.1394
#> [35,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [36,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [37,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#> [38,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#> [39,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#> [40,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#>  [2,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#>  [3,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#>  [4,] 0.4118 0.3641 0.4572 0.0813 0.4367 0.2964 0.4312 0.4155 0.1824 0.1487
#>  [5,] 0.3704 0.3741 0.3839 0.0844 0.3494 0.4380 0.4265 0.2854 0.2808 0.2395
#>  [6,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#>  [7,] 0.5025 0.2829 0.0477 0.0600 0.2811 0.3422 0.5147 0.4372 0.2470 0.1708
#>  [8,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#>  [9,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [10,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [11,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#> [12,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [13,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [14,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [15,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [16,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [17,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [18,] 0.4577 0.3505 0.1845 0.0175 0.1890 0.1967 0.1041 0.0550 0.0492 0.0622
#> [19,] 0.2369 0.1720 0.1878 0.0061 0.3250 0.2575 0.2423 0.2706 0.2323 0.1724
#> [20,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [21,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [22,] 0.1309 0.2408 0.1780 0.1021 0.1598 0.5657 0.6443 0.4241 0.4567 0.5760
#> [23,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [24,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [25,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [26,] 0.1592 0.1626 0.2356 0.0596 0.2483 0.2437 0.2715 0.1184 0.1157 0.1449
#> [27,] 0.1605 0.2061 0.0734 0.0622 0.0202 0.1638 0.1583 0.1830 0.1886 0.1008
#> [28,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#> [29,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [30,] 0.2735 0.3209 0.2724 0.0062 0.1880 0.1552 0.2522 0.2121 0.1801 0.1473
#> [31,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#> [32,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [33,] 0.1781 0.1586 0.3001 0.1497 0.2208 0.1455 0.2895 0.3203 0.1414 0.0629
#> [34,] 0.1670 0.1275 0.1666 0.0494 0.2574 0.2258 0.2777 0.1613 0.1335 0.1976
#> [35,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [36,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [37,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#> [38,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#> [39,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#> [40,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#>  [2,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#>  [3,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#>  [4,] 0.0138 0.1164 0.2052 0.1069 0.0165 0.0199 0.0208 0.0176 0.0197 0.0210
#>  [5,] 0.0369 0.0805 0.0541 0.0177 0.0419 0.0065 0.0222 0.0045 0.0136 0.0113
#>  [6,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#>  [7,] 0.1343 0.0838 0.0755 0.0304 0.1397 0.0074 0.0069 0.0025 0.0103 0.0074
#>  [8,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#>  [9,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [10,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [11,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#> [12,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [13,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [14,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [15,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [16,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [17,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [18,] 0.0505 0.0247 0.0219 0.0102 0.0362 0.0047 0.0019 0.0041 0.0074 0.0030
#> [19,] 0.1457 0.1175 0.0868 0.0392 0.0420 0.0131 0.0092 0.0078 0.0071 0.0081
#> [20,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [21,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [22,] 0.5293 0.3287 0.1283 0.0698 0.0800 0.0334 0.0342 0.0459 0.0277 0.0172
#> [23,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [24,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [25,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [26,] 0.1883 0.1954 0.1492 0.0511 0.0462 0.0155 0.0189 0.0150 0.0060 0.0082
#> [27,] 0.0663 0.0183 0.0404 0.0108 0.0080 0.0143 0.0091 0.0038 0.0096 0.0142
#> [28,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#> [29,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [30,] 0.0681 0.1091 0.0919 0.0397 0.0130 0.0093 0.0076 0.0065 0.0072 0.0108
#> [31,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#> [32,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [33,] 0.0734 0.0805 0.0608 0.0565 0.1284 0.0286 0.0154 0.0154 0.0156 0.0054
#> [34,] 0.1234 0.1554 0.1057 0.0490 0.0988 0.0097 0.0223 0.0121 0.0108 0.0057
#> [35,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [36,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [37,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#> [38,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#> [39,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#> [40,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#>  [2,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#>  [3,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#>  [4,] 0.0141 0.0049 0.0027 0.0162 0.0059 0.0277 0.0021 0.0569 0.2057 0.3887
#>  [5,] 0.0053 0.0165 0.0141 0.0077 0.0246 0.1215 0.0198 0.2002 0.1516 0.0818
#>  [6,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#>  [7,] 0.0123 0.0069 0.0076 0.0073 0.0030 0.1883 0.0138 0.1422 0.1447 0.0487
#>  [8,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#>  [9,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [10,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [11,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#> [12,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [13,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [14,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [15,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [16,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [17,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [18,] 0.0050 0.0048 0.0017 0.0041 0.0086 0.0696 0.0058 0.0873 0.0616 0.1252
#> [19,] 0.0034 0.0064 0.0037 0.0036 0.0012 0.0865 0.0037 0.1182 0.0999 0.1976
#> [20,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [21,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [22,] 0.0087 0.0046 0.0203 0.0130 0.0115 0.0629 0.0015 0.0130 0.0813 0.1761
#> [23,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [24,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [25,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [26,] 0.0091 0.0038 0.0056 0.0056 0.0048 0.0779 0.0024 0.1365 0.0780 0.1038
#> [27,] 0.0190 0.0140 0.0099 0.0092 0.0052 0.0789 0.0075 0.1440 0.1451 0.1789
#> [28,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#> [29,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [30,] 0.0051 0.0102 0.0041 0.0055 0.0050 0.0612 0.0087 0.0895 0.1107 0.0973
#> [31,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#> [32,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [33,] 0.0030 0.0048 0.0087 0.0101 0.0095 0.1165 0.0068 0.1285 0.1684 0.1830
#> [34,] 0.0028 0.0079 0.0034 0.0046 0.0022 0.1425 0.0021 0.1463 0.1219 0.1697
#> [35,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [36,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [37,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> [38,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#> [39,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> [40,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.8614974 0.3886156 0.8755769 0.7850706 0.8235363 0.7222555 0.4215912
#>  [2,] 0.8966269 0.3857354 0.9031531 0.7996597 0.8443523 0.7448090 0.3794913
#>  [3,] 0.8909412 0.4232890 0.9011987 0.8186127 0.8563030 0.7865401 0.4447149
#>  [4,] 0.8349664 0.2952646 0.8552322 0.7458042 0.7457044 0.6098802 0.2999936
#>  [5,] 0.8599989 0.3167962 0.8860353 0.7450955 0.7985160 0.6714640 0.3076873
#>  [6,] 0.8416816 0.2717396 0.8832689 0.7520619 0.7926685 0.7077469 0.3163726
#>  [7,] 0.8776895 0.3470083 0.8894861 0.8085764 0.8167575 0.7455149 0.3479288
#>  [8,] 0.8645421 0.3818334 0.8897964 0.7842827 0.8430744 0.7299221 0.3564330
#>  [9,] 0.9000707 0.4392531 0.9093097 0.8244833 0.8686916 0.7768837 0.4270769
#> [10,] 0.8449282 0.2757189 0.8550305 0.7324273 0.7703972 0.6255885 0.2910159
#> [11,] 0.7883395 0.2697714 0.8329062 0.7791882 0.7207402 0.6925236 0.2688711
#> [12,] 0.8004212 0.2209592 0.8635630 0.7507362 0.7425677 0.7149848 0.2797296
#> [13,] 0.8783895 0.3485529 0.8948257 0.7776214 0.8340513 0.7039014 0.3383161
#> [14,] 0.9048494 0.4051065 0.9106531 0.8347902 0.8529598 0.7729773 0.4041626
#> [15,] 0.8784653 0.3703471 0.8927633 0.7999721 0.8116357 0.7426648 0.3705834
#> [16,] 0.9014890 0.3881924 0.9149757 0.7969930 0.8630447 0.7511616 0.3873662
#> [17,] 0.8654279 0.2966596 0.8981870 0.7638530 0.8162089 0.7102050 0.3282322
#> [18,] 0.9021987 0.4523963 0.9067895 0.8580457 0.8651891 0.8168618 0.4387955
#> [19,] 0.8760588 0.3886267 0.8910804 0.7900030 0.8213999 0.7155417 0.3998278
#> [20,] 0.8586775 0.3184904 0.8895805 0.7563443 0.8025764 0.6904607 0.3363304
#> [21,] 0.7889011 0.2511632 0.8468875 0.6897504 0.7585261 0.6062178 0.3018552
#> [22,] 0.7484448 0.2693993 0.8214301 0.7023849 0.7179530 0.6687274 0.3064533
#> [23,] 0.8787640 0.3476330 0.8926131 0.7730357 0.8363223 0.6937508 0.3626013
#> [24,] 0.8870908 0.2613118 0.9212369 0.8304837 0.8364007 0.7931626 0.2717828
#> [25,] 0.8977522 0.4868487 0.8993204 0.8237308 0.8631344 0.7906791 0.4895145
#> [26,] 0.8440202 0.3788769 0.8689977 0.7870234 0.8078754 0.7435147 0.4008296
#> [27,] 0.8212091 0.4383670 0.8384457 0.7568774 0.7884248 0.6867012 0.3917103
#> [28,] 0.7407405 0.2976816 0.7965273 0.6572001 0.7108501 0.5641013 0.2845247
#> [29,] 0.8403380 0.4848300 0.8436386 0.7734043 0.7912554 0.7348677 0.4989156
#> [30,] 0.8818750 0.4276679 0.8915553 0.8139362 0.8460297 0.7770198 0.4435173
#> [31,] 0.9131036 0.4187974 0.9243363 0.8291756 0.8884851 0.7799131 0.4280494
#> [32,] 0.7834579 0.3873310 0.8166529 0.7623358 0.7404299 0.7450540 0.3571927
#> [33,] 0.7885305 0.3573262 0.7950879 0.7277787 0.7462174 0.6275854 0.3472223
#> [34,] 0.8445888 0.3740973 0.8688259 0.7808891 0.8089989 0.7420311 0.3874851
#> [35,] 0.9118156 0.4180382 0.9240666 0.8224459 0.8825855 0.7804410 0.4310383
#> [36,] 0.7137227 0.2464123 0.7434411 0.6021321 0.6332264 0.5364152 0.2564245
#> [37,] 0.8530197 0.3926280 0.8629827 0.7460524 0.7993530 0.7106112 0.3924326
#> [38,] 0.8193085 0.2878744 0.8254113 0.7125368 0.7202625 0.6365992 0.3160493
#> [39,] 0.7560651 0.3078777 0.8130150 0.7686873 0.7282380 0.7091650 0.3100265
#> [40,] 0.8543801 0.4078311 0.8689294 0.7966741 0.7998386 0.7513398 0.4225695
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4232145 0.5179279 0.6495245
#>  [2,] 0.3989371 0.4814875 0.6063441
#>  [3,] 0.4721231 0.5503864 0.6971926
#>  [4,] 0.2527164 0.3795647 0.5007705
#>  [5,] 0.3160851 0.3875400 0.5252015
#>  [6,] 0.2735040 0.4019955 0.5780215
#>  [7,] 0.3482636 0.4990397 0.6314717
#>  [8,] 0.3926429 0.4604851 0.5980433
#>  [9,] 0.4725277 0.5487185 0.6787616
#> [10,] 0.2989733 0.3719798 0.5051182
#> [11,] 0.2526834 0.4362206 0.6000626
#> [12,] 0.2150035 0.3891544 0.5911793
#> [13,] 0.3651630 0.4321163 0.5628103
#> [14,] 0.4263583 0.5407853 0.6678609
#> [15,] 0.3714971 0.5096626 0.6336140
#> [16,] 0.4186794 0.4847891 0.6235896
#> [17,] 0.2908924 0.4165231 0.5885624
#> [18,] 0.4894700 0.5990454 0.7275708
#> [19,] 0.3877145 0.4758485 0.6002304
#> [20,] 0.3038782 0.4297325 0.5820286
#> [21,] 0.2843460 0.3518784 0.4982973
#> [22,] 0.2674318 0.3984959 0.5827768
#> [23,] 0.3857578 0.4518473 0.5847962
#> [24,] 0.2735255 0.4337553 0.6432271
#> [25,] 0.5380884 0.6029862 0.7140939
#> [26,] 0.4160052 0.5179753 0.6675843
#> [27,] 0.4270427 0.4816967 0.6219551
#> [28,] 0.2896290 0.3222965 0.4557171
#> [29,] 0.5198315 0.5686237 0.6813757
#> [30,] 0.4702442 0.5481242 0.6890239
#> [31,] 0.4701365 0.5025846 0.6475894
#> [32,] 0.3884384 0.5018709 0.6568289
#> [33,] 0.3710537 0.4360823 0.5580368
#> [34,] 0.4061663 0.5095813 0.6597920
#> [35,] 0.4662482 0.5027471 0.6440123
#> [36,] 0.2500241 0.3222787 0.4508218
#> [37,] 0.4070093 0.4759855 0.6004847
#> [38,] 0.3089924 0.4377148 0.5472967
#> [39,] 0.3009925 0.4585559 0.6146097
#> [40,] 0.4301136 0.5158088 0.6571103
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.4268812 0.5731188
#>  [2,] 0.4304639 0.5695361
#>  [3,] 0.4203721 0.5796279
#>  [4,] 0.4475952 0.5524048
#>  [5,] 0.4437689 0.5562311
#>  [6,] 0.4451430 0.5548570
#>  [7,] 0.4322652 0.5677348
#>  [8,] 0.4324031 0.5675969
#>  [9,] 0.4201880 0.5798120
#> [10,] 0.4474049 0.5525951
#> [11,] 0.4428753 0.5571247
#> [12,] 0.4495597 0.5504403
#> [13,] 0.4370572 0.5629428
#> [14,] 0.4236432 0.5763568
#> [15,] 0.4300549 0.5699451
#> [16,] 0.4291972 0.5708028
#> [17,] 0.4423092 0.5576908
#> [18,] 0.4147985 0.5852015
#> [19,] 0.4311848 0.5688152
#> [20,] 0.4405676 0.5594324
#> [21,] 0.4511747 0.5488253
#> [22,] 0.4463320 0.5536680
#> [23,] 0.4349488 0.5650512
#> [24,] 0.4409167 0.5590833
#> [25,] 0.4124708 0.5875292
#> [26,] 0.4270262 0.5729738
#> [27,] 0.4273210 0.5726790
#> [28,] 0.4518640 0.5481360
#> [29,] 0.4162295 0.5837705
#> [30,] 0.4206132 0.5793868
#> [31,] 0.4238986 0.5761014
#> [32,] 0.4289880 0.5710120
#> [33,] 0.4369553 0.5630447
#> [34,] 0.4283402 0.5716598
#> [35,] 0.4243108 0.5756892
#> [36,] 0.4565770 0.5434230
#> [37,] 0.4312717 0.5687283
#> [38,] 0.4428157 0.5571843
#> [39,] 0.4384617 0.5615383
#> [40,] 0.4253013 0.5746987
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]        [,2]     [,3]      [,4]      [,5]      [,6]         [,7]
#>  [1,] 1.8277831 -0.45313503 1.951195 1.2954639 1.5404922 0.9556776 -0.316244767
#>  [2,] 2.1602954 -0.46527393 2.232761 1.3841687 1.6909748 1.0711156 -0.491707897
#>  [3,] 2.1003913 -0.30928596 2.210616 1.5069764 1.7849172 1.3041947 -0.222048411
#>  [4,] 1.6212422 -0.86995057 1.776242 1.0763584 1.0758319 0.4468086 -0.847328149
#>  [5,] 1.8152805 -0.76853445 2.050868 1.0726235 1.3770449 0.7148138 -0.810953575
#>  [6,] 1.6707933 -0.98581451 2.023757 1.1096395 1.3410860 0.8844663 -0.770492435
#>  [7,] 1.9707296 -0.63221579 2.085502 1.4407866 1.4945319 1.0748331 -0.628155992
#>  [8,] 1.8535389 -0.48177351 2.088663 1.2908009 1.6812831 0.9942274 -0.590879682
#>  [9,] 2.1980109 -0.24419390 2.305235 1.5470230 1.8894390 1.2475973 -0.293787264
#> [10,] 1.6953631 -0.96579789 1.774614 1.0069729 1.2105552 0.5133372 -0.890455388
#> [11,] 1.3149452 -0.99578281 1.606366 1.2609421 0.9481362 0.8119440 -1.000357522
#> [12,] 1.3889287 -1.26008556 1.845204 1.1025428 1.0593575 0.9197188 -0.945803260
#> [13,] 1.9772669 -0.62540631 2.141010 1.2518582 1.6146162 0.8659456 -0.670807479
#> [14,] 2.2523078 -0.38423228 2.321636 1.6199639 1.7580062 1.2251999 -0.388150455
#> [15,] 1.9779770 -0.53072818 2.119283 1.3861201 1.4606738 1.0598652 -0.529714930
#> [16,] 2.2138796 -0.45491676 2.375960 1.3676052 1.8408115 1.1048173 -0.458396739
#> [17,] 1.8611241 -0.86325554 2.177241 1.1739208 1.4908707 0.8963799 -0.716191291
#> [18,] 2.2218967 -0.19099314 2.275050 1.7991525 1.8590750 1.4952289 -0.246051816
#> [19,] 1.9556257 -0.45308863 2.101824 1.3249436 1.5258612 0.9224533 -0.406182855
#> [20,] 1.8043489 -0.76071819 2.086463 1.1327407 1.4024754 0.8022741 -0.679690415
#> [21,] 1.3183142 -1.09241811 1.710395 0.7989529 1.1446159 0.4314413 -0.838478890
#> [22,] 1.0903352 -0.99767248 1.526066 0.8586806 0.9343303 0.7024348 -0.816753205
#> [23,] 1.9807773 -0.62946003 2.117715 1.2255325 1.6311148 0.8177135 -0.564091330
#> [24,] 2.0613632 -1.03916159 2.459272 1.5890593 1.6316877 1.3440954 -0.985596173
#> [25,] 2.1724946 -0.05261726 2.189696 1.5418316 1.8415706 1.3290234 -0.041948251
#> [26,] 1.6884499 -0.49431809 1.892126 1.3070756 1.4362639 1.0643172 -0.402009485
#> [27,] 1.5245608 -0.24779211 1.646708 1.1356353 1.3154568 0.7847417 -0.440128872
#> [28,] 1.0498208 -0.85836261 1.364730 0.6508416 0.8995163 0.2578241 -0.922126883
#> [29,] 1.6607452 -0.06069871 1.685554 1.2276347 1.3325090 1.0194614 -0.004337592
#> [30,] 2.0103072 -0.29137228 2.106727 1.4757922 1.7037945 1.2483830 -0.226899458
#> [31,] 2.3521322 -0.32771204 2.502777 1.5797958 2.0753598 1.2651600 -0.289813876
#> [32,] 1.2859323 -0.45854520 1.493833 1.1655284 1.0482043 1.0724047 -0.587569497
#> [33,] 1.3160903 -0.58698809 1.355871 0.9833817 1.0785388 0.5218718 -0.631271495
#> [34,] 1.6927750 -0.51467891 1.890617 1.2708553 1.4435183 1.0565522 -0.457896034
#> [35,] 2.3360083 -0.33083217 2.498927 1.5330074 2.0171451 1.2682379 -0.277616089
#> [36,] 0.9135335 -1.11783925 1.063931 0.4143569 0.5460832 0.1459192 -1.064636113
#> [37,] 1.7584844 -0.43627885 1.840287 1.0776679 1.3822555 0.8983543 -0.437098507
#> [38,] 1.5116694 -0.90572993 1.553449 0.9077369 0.9457642 0.5606340 -0.771987642
#> [39,] 1.1312263 -0.81005998 1.469721 1.2009138 0.9857010 0.8913320 -0.799995631
#> [40,] 1.7693761 -0.37293870 1.891526 1.3656357 1.3852862 1.1057706 -0.312234156
#>              [,8]         [,9]        [,10]
#>  [1,] -0.30959137  0.071742222  0.616949690
#>  [2,] -0.40989598 -0.074083696  0.431970631
#>  [3,] -0.11162320  0.202232087  0.833964818
#>  [4,] -1.08417676 -0.491396121  0.003081831
#>  [5,] -0.77182212 -0.457664671  0.100891588
#>  [6,] -0.97691654 -0.397157345  0.314656869
#>  [7,] -0.62668032 -0.003841028  0.538535592
#>  [8,] -0.43621632 -0.158389740  0.397318745
#>  [9,] -0.10999988  0.195494433  0.748086617
#> [10,] -0.85219151 -0.523732855  0.020473349
#> [11,] -1.08435167 -0.256515152  0.405725925
#> [12,] -1.29502488 -0.450868074  0.368842803
#> [13,] -0.55302452 -0.273221679  0.252575534
#> [14,] -0.29672475  0.163504512  0.698526189
#> [15,] -0.52579956  0.038655058  0.547752604
#> [16,] -0.32819678 -0.060862532  0.504812211
#> [17,] -0.89105386 -0.337062930  0.358025772
#> [18,] -0.04212643  0.401489203  0.982332464
#> [19,] -0.45692938 -0.096681218  0.406425105
#> [20,] -0.82889765 -0.282942777  0.331106480
#> [21,] -0.92300483 -0.610792687 -0.006810985
#> [22,] -1.00769188 -0.411736054  0.334183022
#> [23,] -0.46517957 -0.193209660  0.342493739
#> [24,] -0.97680858 -0.266545760  0.589398374
#> [25,]  0.15264950  0.417923162  0.915351147
#> [26,] -0.33919420  0.071932203  0.697279602
#> [27,] -0.29392706 -0.073245828  0.497855026
#> [28,] -0.89718667 -0.743237854 -0.177596840
#> [29,]  0.07936749  0.276238156  0.760101273
#> [30,] -0.11916400  0.193094604  0.795559952
#> [31,] -0.11959619  0.010338663  0.608459867
#> [32,] -0.45388104  0.007483657  0.649194339
#> [33,] -0.52769926 -0.257077241  0.233198201
#> [34,] -0.37983665  0.038329747  0.662367321
#> [35,] -0.13521279  0.010988562  0.592821778
#> [36,] -1.09848396 -0.743319379 -0.197350856
#> [37,] -0.37634253 -0.096131898  0.407485053
#> [38,] -0.80483400 -0.250441723  0.189754203
#> [39,] -0.84257604 -0.166157671  0.466730954
#> [40,] -0.28138783  0.063256187  0.650443026
#> 
#> $pre[[3]]
#>              [,1]      [,2]
#>  [1,] -0.08917020 0.2054172
#>  [2,] -0.08306190 0.1968967
#>  [3,] -0.10289699 0.2183492
#>  [4,] -0.04847625 0.1619158
#>  [5,] -0.05618549 0.1696943
#>  [6,] -0.05017378 0.1701412
#>  [7,] -0.07740875 0.1952065
#>  [8,] -0.08076381 0.1912895
#>  [9,] -0.10416656 0.2178351
#> [10,] -0.04786432 0.1632971
#> [11,] -0.05876182 0.1707389
#> [12,] -0.04150580 0.1609441
#> [13,] -0.06993533 0.1831785
#> [14,] -0.09583510 0.2120004
#> [15,] -0.08243686 0.1991901
#> [16,] -0.08457242 0.2005549
#> [17,] -0.05507484 0.1767206
#> [18,] -0.11542415 0.2287391
#> [19,] -0.08219856 0.1948202
#> [20,] -0.05954794 0.1793109
#> [21,] -0.04056329 0.1553621
#> [22,] -0.05143629 0.1640660
#> [23,] -0.07237693 0.1893109
#> [24,] -0.05742835 0.1800143
#> [25,] -0.12082147 0.2329390
#> [26,] -0.08976605 0.2042287
#> [27,] -0.09388389 0.1989058
#> [28,] -0.04552339 0.1476189
#> [29,] -0.11601725 0.2222541
#> [30,] -0.10330547 0.2169511
#> [31,] -0.09733921 0.2094502
#> [32,] -0.09089053 0.1950907
#> [33,] -0.07388074 0.1796475
#> [34,] -0.08688448 0.2017417
#> [35,] -0.09636237 0.2087392
#> [36,] -0.03314858 0.1409822
#> [37,] -0.08226755 0.1943971
#> [38,] -0.05607422 0.1736682
#> [39,] -0.07083380 0.1765739
#> [40,] -0.09551960 0.2055283
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.5731188 -0.5731188
#>  [2,]  0.5695361 -0.5695361
#>  [3,] -0.4203721  0.4203721
#>  [4,]  0.5524048 -0.5524048
#>  [5,]  0.5562311 -0.5562311
#>  [6,]  0.5548570 -0.5548570
#>  [7,] -0.4322652  0.4322652
#>  [8,]  0.5675969 -0.5675969
#>  [9,]  0.5798120 -0.5798120
#> [10,] -0.4474049  0.4474049
#> [11,] -0.4428753  0.4428753
#> [12,]  0.5504403 -0.5504403
#> [13,]  0.5629428 -0.5629428
#> [14,]  0.5763568 -0.5763568
#> [15,] -0.4300549  0.4300549
#> [16,]  0.5708028 -0.5708028
#> [17,]  0.5576908 -0.5576908
#> [18,] -0.4147985  0.4147985
#> [19,]  0.5688152 -0.5688152
#> [20,]  0.5594324 -0.5594324
#> [21,] -0.4511747  0.4511747
#> [22,]  0.5536680 -0.5536680
#> [23,]  0.5650512 -0.5650512
#> [24,] -0.4409167  0.4409167
#> [25,] -0.4124708  0.4124708
#> [26,]  0.5729738 -0.5729738
#> [27,] -0.4273210  0.4273210
#> [28,]  0.5481360 -0.5481360
#> [29,] -0.4162295  0.4162295
#> [30,] -0.4206132  0.4206132
#> [31,] -0.4238986  0.4238986
#> [32,] -0.4289880  0.4289880
#> [33,]  0.5630447 -0.5630447
#> [34,]  0.5716598 -0.5716598
#> [35,] -0.4243108  0.4243108
#> [36,]  0.5434230 -0.5434230
#> [37,]  0.5687283 -0.5687283
#> [38,]  0.5571843 -0.5571843
#> [39,] -0.4384617  0.4384617
#> [40,] -0.4253013  0.4253013
#> 
#> $L
#> [1] 0.6854340 0.7701846 0.7882947 0.6745578 0.7797565 0.7142745
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536