Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11          V12         V13          V14
#>  [1,] -0.2068574 -0.03262733 -0.15859007 -0.009855225 -0.10642937  0.027806774
#>  [2,] -0.3284052 -0.15320028  0.01141172 -0.120622297 -0.05767054  0.004376619
#>  [3,] -0.4328552 -0.13358863 -0.13405279 -0.085116948  0.01248790 -0.072288599
#>  [4,] -0.2015676 -0.01585579 -0.02839412 -0.111374641 -0.14134162 -0.220406281
#>  [5,] -0.2475186 -0.19884616 -0.03912956 -0.044986559 -0.17573786 -0.099000928
#>  [6,] -0.3224862 -0.25818811 -0.17298162 -0.134363711 -0.13979527 -0.128378007
#>  [7,] -0.3687111 -0.07080518 -0.11463903 -0.130491719  0.01032931 -0.026304649
#>  [8,] -0.3944246 -0.19505623 -0.16462106 -0.189819168 -0.13445381 -0.102109410
#>  [9,] -0.3108840 -0.10465443 -0.13549610 -0.091044204  0.02005695  0.040066376
#> [10,] -0.2630510 -0.03948121 -0.09685425 -0.069387457 -0.05234309 -0.050192889
#>                V15          V16          V17          V18         V19
#>  [1,] -0.100817423 -0.071942056  0.012441659 -0.012993718 -0.01515677
#>  [2,] -0.031614017 -0.073962193  0.118022238 -0.010475829  0.06275953
#>  [3,] -0.096147589  0.074694967  0.030247820  0.054832803  0.11871831
#>  [4,] -0.084936170 -0.161920035 -0.169726991 -0.090836218 -0.06028319
#>  [5,] -0.001352922 -0.114088897 -0.074576382  0.080856108  0.08240706
#>  [6,] -0.057475489 -0.006144976 -0.004158116 -0.078983541 -0.07477970
#>  [7,]  0.044017087 -0.017760622  0.043462940  0.083106747  0.11609179
#>  [8,] -0.074808566 -0.098680476 -0.127592379 -0.076928870 -0.19413980
#>  [9,] -0.014492606  0.119366957  0.104026939  0.079864075  0.11634455
#> [10,] -0.019340687 -0.024006174  0.017437580 -0.001998181  0.11066758
#>               V2          V20         V21         V22         V23          V24
#>  [1,] -0.2642693  0.097576248  0.15968796  0.09502534  0.01368218  0.023281610
#>  [2,] -0.1989439  0.111692040  0.17244611  0.09803351  0.18525180  0.149680031
#>  [3,] -0.3385971  0.101058756  0.19338651  0.26311800  0.15480361  0.234284486
#>  [4,] -0.3344391 -0.008201524 -0.08407769  0.04893808 -0.02247641 -0.027267335
#>  [5,] -0.2401441  0.054134073  0.06797060  0.11899712  0.03257326  0.105638396
#>  [6,] -0.3044322  0.079552477  0.01789374 -0.08348462  0.11984975 -0.024901108
#>  [7,] -0.2640405  0.038022498  0.08252933  0.13301266  0.09617093  0.117057798
#>  [8,] -0.2999307 -0.164094105 -0.04691045  0.03117370 -0.06023691 -0.006059536
#>  [9,] -0.3732050  0.207411545  0.02452120  0.15141389  0.10759674  0.165344458
#> [10,] -0.3555603  0.091505687  0.20052050  0.01118027  0.13205292  0.150736276
#>               V25         V26          V27          V28          V29         V3
#>  [1,] -0.02229309  0.03867345  0.070680357  0.137349433  0.009068694 -0.2769330
#>  [2,]  0.22582125  0.06131824  0.250483887  0.258299428  0.145259303 -0.2468587
#>  [3,]  0.15480409  0.21979096  0.257111767  0.149958757  0.246978344 -0.3717582
#>  [4,]  0.09021786  0.12236001 -0.033506360  0.025060695  0.071369319 -0.1724224
#>  [5,]  0.11461990  0.13873818  0.167912046  0.162155700  0.155337300 -0.1692105
#>  [6,]  0.02334977  0.03028792  0.135842811 -0.046749105 -0.053259786 -0.3443230
#>  [7,]  0.10717952  0.18971417  0.104073771  0.210763929 -0.030986053 -0.3071285
#>  [8,]  0.03991751 -0.04141037  0.106896875 -0.009514131  0.144365453 -0.3553457
#>  [9,]  0.16476211  0.13629331  0.004248365  0.019829079  0.083582863 -0.3132048
#> [10,]  0.18448585  0.21211935  0.275941763  0.287595284  0.244557353 -0.3224331
#>               V30          V31         V32         V33         V34         V35
#>  [1,]  0.15724158  0.043834499  0.09652997  0.06668598 -0.02099581  0.02977611
#>  [2,]  0.07621525 -0.086970228  0.02684339 -0.06872979 -0.07558785 -0.09014471
#>  [3,]  0.18496096  0.137336295  0.03624435  0.15934278  0.10077558  0.02849606
#>  [4,]  0.07065563 -0.002574991 -0.05764991 -0.02138659  0.02761408 -0.07329557
#>  [5,] -0.01059038 -0.041628806 -0.05341932 -0.07582608 -0.08347979 -0.06413390
#>  [6,] -0.12951853 -0.062880360 -0.17379735 -0.03010826 -0.13971995 -0.16667713
#>  [7,]  0.06288182  0.024210824 -0.11263252 -0.13429594 -0.04864710 -0.15880140
#>  [8,]  0.16964538  0.075699957 -0.03855562  0.02704012  0.05977246  0.04385515
#>  [9,]  0.04690896 -0.025783086 -0.05110830 -0.13534056 -0.08742660  0.04618302
#> [10,]  0.10954921  0.092508871  0.03271086 -0.03770029 -0.09802429 -0.02454225
#>               V36         V37          V38          V39         V4          V40
#>  [1,] -0.01050308  0.06459876  0.089797657 -0.069607097 -0.2240619 -0.061181556
#>  [2,] -0.06210937  0.01144293 -0.142662530 -0.015907144 -0.2379859 -0.141062036
#>  [3,] -0.08389387  0.03829555  0.085738604 -0.004733582 -0.3654578  0.057624293
#>  [4,] -0.04629266 -0.07428728  0.012104977 -0.125364597 -0.1987695 -0.060873196
#>  [5,] -0.02087045 -0.10613206 -0.183598499 -0.122451128 -0.2636554 -0.135392093
#>  [6,] -0.18140800 -0.23455516 -0.087650549 -0.140246496 -0.2432603 -0.073348704
#>  [7,] -0.09313139 -0.09729936 -0.108889456 -0.017962942 -0.2263338  0.026196567
#>  [8,]  0.14750765  0.11288057  0.008711464 -0.062628979 -0.2115259 -0.083679114
#>  [9,] -0.03152882  0.03657830 -0.021660004 -0.039464842 -0.3629528 -0.138522464
#> [10,] -0.07938641 -0.10369088  0.018321798  0.019384640 -0.2202142  0.009818586
#>                V41         V42         V43         V44          V45         V46
#>  [1,] -0.062297210 -0.11357819 -0.07334160 -0.07234782 -0.006865763 -0.21698501
#>  [2,] -0.147171113 -0.17820651 -0.15103506 -0.03406761 -0.067581652 -0.14413327
#>  [3,]  0.048169771 -0.05482488 -0.01988295 -0.19073905 -0.127917183 -0.21550870
#>  [4,] -0.035711123 -0.09503877 -0.06372008 -0.13216277 -0.172588452 -0.22611105
#>  [5,] -0.190389783 -0.20123055 -0.14625800 -0.13702414 -0.169734302 -0.13718998
#>  [6,] -0.224428989 -0.07530532 -0.13841004 -0.11439691 -0.173153179 -0.29251713
#>  [7,] -0.072528190  0.01718077 -0.10999104 -0.15167694 -0.086806617 -0.08253348
#>  [8,] -0.014215625 -0.13773663 -0.08405875 -0.05228805 -0.181116955 -0.12278940
#>  [9,] -0.003724759 -0.16187358 -0.15487518 -0.16601717 -0.206809982 -0.15270370
#> [10,] -0.100493310 -0.02569160 -0.14410186 -0.04386301 -0.111235116 -0.03773676
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.2159623 -0.1060655 -0.3089086 -0.2717747 -0.3022154 -0.2731709
#>  [2,] -0.2024936 -0.1672352 -0.1921950 -0.1335445 -0.4005412 -0.2508543
#>  [3,] -0.1805536 -0.2726876 -0.2839164 -0.2168977 -0.3191354 -0.4560107
#>  [4,] -0.1146837 -0.2545551 -0.2699126 -0.1394953 -0.3811302 -0.3611985
#>  [5,] -0.1708117 -0.2464181 -0.2524791 -0.3221528 -0.3479881 -0.2482761
#>  [6,] -0.1416053 -0.2349756 -0.2544021 -0.1925895 -0.3097069 -0.2554324
#>  [7,] -0.1321900 -0.2092291 -0.2801181 -0.2331986 -0.2580414 -0.2553037
#>  [8,] -0.2305852 -0.2775586 -0.3010397 -0.2994003 -0.3464675 -0.3798846
#>  [9,] -0.1746144 -0.2817551 -0.3454944 -0.2561657 -0.3287994 -0.4368526
#> [10,] -0.2119253 -0.2051994 -0.2034378 -0.2322138 -0.2877647 -0.2494443
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3605312 -0.2445291 -0.3658584 -0.2776438 -0.3201132 -0.2246212
#>  [2,] -0.3415991 -0.2645084 -0.3569729 -0.3670476 -0.3930720 -0.3083038
#>  [3,] -0.3014283 -0.4018765 -0.4251792 -0.3916662 -0.4047036 -0.4587888
#>  [4,] -0.3253981 -0.2831620 -0.3253783 -0.2332011 -0.3309316 -0.2469986
#>  [5,] -0.3520598 -0.4199018 -0.2727932 -0.3811157 -0.2445473 -0.3764296
#>  [6,] -0.3459033 -0.2900033 -0.2849137 -0.3297192 -0.3192082 -0.3910210
#>  [7,] -0.2936813 -0.3045611 -0.2169734 -0.3171088 -0.3874093 -0.2685043
#>  [8,] -0.3647604 -0.2566099 -0.4216205 -0.3708256 -0.2833358 -0.3598652
#>  [9,] -0.3296299 -0.2915277 -0.3399842 -0.3917041 -0.3359687 -0.4379100
#> [10,] -0.3398640 -0.3951602 -0.2332093 -0.3433573 -0.3903010 -0.3427849
#>              V58        V59         V6        V60         V7         V8
#>  [1,] -0.3480669 -0.3340040 -0.2916827 -0.3205916 -0.1702919 -0.1492568
#>  [2,] -0.2402862 -0.3017092 -0.2321582 -0.4013088 -0.2378603 -0.2123136
#>  [3,] -0.4764511 -0.4048320 -0.1855617 -0.4522337 -0.2748679 -0.1922982
#>  [4,] -0.3682662 -0.3969079 -0.1333332 -0.3708323 -0.1775133 -0.2089119
#>  [5,] -0.3268283 -0.2956452 -0.1961228 -0.3128477 -0.1418622 -0.1258344
#>  [6,] -0.4012675 -0.2568270 -0.2806480 -0.3194987 -0.3013004 -0.2067883
#>  [7,] -0.2552257 -0.2474453 -0.2798600 -0.2438039 -0.1489102 -0.2133810
#>  [8,] -0.2793347 -0.4312592 -0.1805493 -0.4300204 -0.1653587 -0.1243962
#>  [9,] -0.2803149 -0.3324832 -0.1114813 -0.3294976 -0.1255234 -0.1324122
#> [10,] -0.3499449 -0.3142955 -0.2409942 -0.2143608 -0.1356980 -0.2577929
#>                V9
#>  [1,] -0.18718536
#>  [2,] -0.16541736
#>  [3,] -0.06616512
#>  [4,] -0.10916009
#>  [5,] -0.15555268
#>  [6,] -0.10710938
#>  [7,] -0.14517231
#>  [8,] -0.20835428
#>  [9,] -0.18052668
#> [10,] -0.02499961
#> 
#> $W[[2]]
#>              [,1]        [,2]        [,3]        [,4]         [,5]        [,6]
#> [1,] -0.025950869  0.03038646 -0.05675894 -0.09021608 -0.005122684 -0.00956330
#> [2,]  0.003236862 -0.04048033  0.06017486  0.03149682  0.099797813 -0.02247927
#>             [,7]        [,8]        [,9]       [,10]
#> [1,] -0.01397911 -0.08079647 -0.03144533  0.03750032
#> [2,] -0.04810950  0.19379560 -0.00114653 -0.04027348
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1          V10           V11           V12          V13
#>  [1,]  8.500964e-05 7.337500e-04  8.161291e-04  1.053158e-03 0.0011989113
#>  [2,]  3.473385e-06 1.645781e-04  8.889953e-05  2.703369e-04 0.0005383363
#>  [3,]  1.078380e-04 7.627390e-04  8.673978e-04  1.031916e-03 0.0010164609
#>  [4,]  1.415939e-04 1.141967e-03  1.426390e-03  1.694974e-03 0.0016016932
#>  [5,]  1.716104e-04 1.332639e-03  1.642716e-03  1.958623e-03 0.0019057241
#>  [6,]  4.170398e-06 5.509773e-05 -2.973625e-05 -1.499937e-05 0.0001470734
#>  [7,]  3.624466e-05 3.674252e-04  3.443302e-04  5.300325e-04 0.0007602086
#>  [8,]  2.854039e-04 2.238961e-03  2.839833e-03  3.311761e-03 0.0029630677
#>  [9,]  7.843148e-05 6.574907e-04  7.392500e-04  9.853802e-04 0.0011175750
#> [10,] -7.704731e-07 1.599664e-04  8.776079e-05  2.596728e-04 0.0005137272
#>                V14          V15          V16         V17          V18
#>  [1,] 0.0012876688 0.0013194356 0.0016260099 0.001969081 0.0022558444
#>  [2,] 0.0008897838 0.0011634009 0.0015319930 0.001775861 0.0018840188
#>  [3,] 0.0008888431 0.0007343416 0.0008348721 0.001031569 0.0012966062
#>  [4,] 0.0012997547 0.0009875104 0.0010934109 0.001461297 0.0019058393
#>  [5,] 0.0016017221 0.0012610307 0.0014135309 0.001836654 0.0023425873
#>  [6,] 0.0003755633 0.0005443546 0.0007278256 0.000793098 0.0007986735
#>  [7,] 0.0010285940 0.0012185395 0.0015690532 0.001843128 0.0020055733
#>  [8,] 0.0021258769 0.0012991364 0.0012980915 0.001884456 0.0027374049
#>  [9,] 0.0011704700 0.0011473888 0.0014193731 0.001729022 0.0020021727
#> [10,] 0.0008354936 0.0010845028 0.0014390202 0.001687918 0.0017855061
#>                V19           V2          V20          V21          V22
#>  [1,] 0.0025924351 1.103527e-04 0.0026812718 0.0027925555 0.0028279851
#>  [2,] 0.0018288469 3.982692e-05 0.0013727898 0.0013551092 0.0016398241
#>  [3,] 0.0016899068 1.167645e-04 0.0020981745 0.0022648054 0.0020793066
#>  [4,] 0.0025734254 1.512476e-04 0.0031783396 0.0033868459 0.0032090346
#>  [5,] 0.0031169955 1.882931e-04 0.0038150716 0.0040505076 0.0038288251
#>  [6,] 0.0007913235 1.157281e-05 0.0006896628 0.0006815681 0.0007418232
#>  [7,] 0.0020924049 6.655117e-05 0.0018646706 0.0018951803 0.0020653195
#>  [8,] 0.0041379679 2.888945e-04 0.0056244285 0.0060681319 0.0055498004
#>  [9,] 0.0023181744 1.056782e-04 0.0023805617 0.0025008332 0.0025381009
#> [10,] 0.0017252207 3.583855e-05 0.0012692683 0.0012650261 0.0015756427
#>                V23          V24         V25         V26         V27         V28
#>  [1,] 0.0028881766 0.0029676776 0.002934813 0.002909503 0.002917628 0.002805240
#>  [2,] 0.0019816788 0.0021769173 0.002220675 0.002173002 0.002120290 0.002067355
#>  [3,] 0.0019107922 0.0018909809 0.001818409 0.001841200 0.001878196 0.001829035
#>  [4,] 0.0029461864 0.0028776844 0.002761618 0.002750704 0.002819823 0.002677190
#>  [5,] 0.0035510459 0.0034781125 0.003337946 0.003338770 0.003418611 0.003266330
#>  [6,] 0.0009000076 0.0009918778 0.001050988 0.001082655 0.001063240 0.001026117
#>  [7,] 0.0023038193 0.0024611607 0.002488720 0.002449005 0.002423452 0.002364055
#>  [8,] 0.0048386986 0.0046115603 0.004325615 0.004347769 0.004529605 0.004320491
#>  [9,] 0.0026020295 0.0026980077 0.002665395 0.002638402 0.002655785 0.002583296
#> [10,] 0.0019129986 0.0020836832 0.002107601 0.002069791 0.002022036 0.001964950
#>                V29           V3         V30          V31          V32
#>  [1,] 0.0024440649 1.381015e-04 0.002023426 0.0018234815 0.0015722951
#>  [2,] 0.0020676694 8.045701e-05 0.002008902 0.0020676309 0.0014472867
#>  [3,] 0.0014854082 1.187611e-04 0.001151649 0.0009153778 0.0010569547
#>  [4,] 0.0020425214 1.543120e-04 0.001307769 0.0008902735 0.0011247990
#>  [5,] 0.0025626099 1.952927e-04 0.001768594 0.0013117565 0.0015594807
#>  [6,] 0.0009709809 2.439603e-05 0.000916807 0.0009153345 0.0006383195
#>  [7,] 0.0022335975 9.972470e-05 0.002042717 0.0019945939 0.0015188716
#>  [8,] 0.0031004845 2.746265e-04 0.001677453 0.0008170614 0.0016369240
#>  [9,] 0.0022758717 1.285665e-04 0.001884054 0.0017400146 0.0015280750
#> [10,] 0.0019550954 7.697057e-05 0.001883759 0.0019617573 0.0013755936
#>                V33          V34          V35           V36          V37
#>  [1,] 0.0015328897 0.0016570686 0.0016641051  0.0016456054 0.0017776018
#>  [2,] 0.0016060724 0.0018450627 0.0020036871  0.0021851003 0.0021240428
#>  [3,] 0.0009551727 0.0009557342 0.0008701694  0.0007342202 0.0008995664
#>  [4,] 0.0008115564 0.0007232413 0.0005835851  0.0004122298 0.0007453758
#>  [5,] 0.0012644995 0.0011789862 0.0009823001  0.0007661992 0.0011119710
#>  [6,] 0.0007512807 0.0009026746 0.0009710479  0.0009995720 0.0009381046
#>  [7,] 0.0016209740 0.0018314953 0.0019427176  0.0020292754 0.0020218839
#>  [8,] 0.0009281469 0.0006031534 0.0001905067 -0.0002454423 0.0004379281
#>  [9,] 0.0014926325 0.0016026754 0.0016253722  0.0016188052 0.0017284133
#> [10,] 0.0015011564 0.0017077766 0.0018496675  0.0020501428 0.0020098030
#>                V38          V39           V4          V40          V41
#>  [1,] 0.0018207971 0.0017091343 1.826328e-04 0.0015888179 0.0013534869
#>  [2,] 0.0017850564 0.0014537537 3.714013e-05 0.0014490885 0.0011680871
#>  [3,] 0.0011846044 0.0012260316 1.931717e-04 0.0010585156 0.0009627003
#>  [4,] 0.0012051155 0.0014283908 2.862269e-04 0.0012623869 0.0011121879
#>  [5,] 0.0016376379 0.0018625877 3.379636e-04 0.0016180950 0.0014384472
#>  [6,] 0.0007656037 0.0005733498 8.832116e-06 0.0005337774 0.0004277980
#>  [7,] 0.0018064793 0.0015302804 9.020071e-05 0.0014633604 0.0011984633
#>  [8,] 0.0015854632 0.0022365425 5.656472e-04 0.0018750471 0.0017333665
#>  [9,] 0.0017735357 0.0016840017 1.619937e-04 0.0015672976 0.0013207663
#> [10,] 0.0016798665 0.0013749534 3.407274e-05 0.0013977613 0.0011547255
#>                V42          V43          V44           V45          V46
#>  [1,] 0.0012010161 0.0009445881 8.060556e-04  7.225781e-04 5.899192e-04
#>  [2,] 0.0007072495 0.0003962079 2.389119e-04  5.547956e-05 1.149683e-04
#>  [3,] 0.0010338789 0.0008961815 8.274780e-04  8.675329e-04 6.754244e-04
#>  [4,] 0.0013853667 0.0013019650 1.239210e-03  1.250820e-03 9.456469e-04
#>  [5,] 0.0017028597 0.0015563918 1.473737e-03  1.485399e-03 1.119807e-03
#>  [6,] 0.0002310831 0.0001040485 4.128155e-05 -1.936623e-06 2.013685e-05
#>  [7,] 0.0008537644 0.0005732734 4.306315e-04  2.932745e-04 2.824851e-04
#>  [8,] 0.0025088831 0.0024855850 2.430104e-03  2.550648e-03 1.891324e-03
#>  [9,] 0.0011754204 0.0009261683 8.018774e-04  7.202420e-04 5.808532e-04
#> [10,] 0.0007155675 0.0004038576 2.531755e-04  7.967097e-05 1.305884e-04
#>                V47          V48          V49           V5          V50
#>  [1,] 4.757263e-04 0.0003289953 1.661440e-04 3.415275e-04 7.472552e-05
#>  [2,] 2.166590e-04 0.0001265326 6.617667e-05 1.315257e-04 4.900832e-05
#>  [3,] 4.404227e-04 0.0003248935 1.659563e-04 3.032400e-04 6.302763e-05
#>  [4,] 6.521326e-04 0.0004693799 2.390816e-04 4.801796e-04 7.539906e-05
#>  [5,] 7.575681e-04 0.0005519679 2.793282e-04 5.563703e-04 9.463893e-05
#>  [6,] 4.667513e-05 0.0000318448 1.303970e-05 3.864411e-05 2.136526e-05
#>  [7,] 2.970050e-04 0.0001937173 9.724494e-05 1.989084e-04 5.793828e-05
#>  [8,] 1.215914e-03 0.0008853251 4.476905e-04 9.136661e-04 1.289950e-04
#>  [9,] 4.625170e-04 0.0003231560 1.620351e-04 3.100172e-04 6.890280e-05
#> [10,] 2.302866e-04 0.0001392986 7.618226e-05 1.327335e-04 4.887426e-05
#>                V51          V52          V53          V54          V55
#>  [1,] 6.377575e-05 4.961380e-05 4.154187e-05 4.372173e-05 4.668415e-05
#>  [2,] 1.100554e-05 1.352817e-05 2.206639e-05 1.800677e-05 3.456346e-05
#>  [3,] 6.852610e-05 5.009116e-05 3.522959e-05 4.251809e-05 3.577670e-05
#>  [4,] 1.071080e-04 7.366463e-05 4.891945e-05 5.503911e-05 3.951437e-05
#>  [5,] 1.225842e-04 8.825710e-05 6.008001e-05 6.814035e-05 5.387950e-05
#>  [6,] 6.906413e-07 5.782037e-06 9.645165e-06 8.855928e-06 1.747921e-05
#>  [7,] 2.905636e-05 2.691431e-05 2.972654e-05 2.800310e-05 3.975931e-05
#>  [8,] 2.131199e-04 1.432402e-04 8.875145e-05 1.034937e-04 6.262859e-05
#>  [9,] 6.072123e-05 4.535736e-05 3.857771e-05 4.033482e-05 4.334419e-05
#> [10,] 1.260440e-05 1.288672e-05 2.060486e-05 1.614509e-05 3.202662e-05
#>                V56          V57          V58          V59           V6
#>  [1,] 3.726436e-05 2.865061e-05 3.432775e-05 3.598655e-05 0.0004928165
#>  [2,] 1.920215e-05 2.922170e-05 1.792708e-05 1.947364e-05 0.0003331149
#>  [3,] 3.242339e-05 1.644104e-05 3.129706e-05 3.500781e-05 0.0003488025
#>  [4,] 4.396286e-05 1.784657e-05 4.145584e-05 4.069024e-05 0.0005172917
#>  [5,] 5.419636e-05 2.402984e-05 5.012476e-05 5.138535e-05 0.0006188374
#>  [6,] 7.615677e-06 1.210994e-05 5.385876e-06 6.735184e-06 0.0001283990
#>  [7,] 2.616748e-05 2.906608e-05 2.358824e-05 2.551175e-05 0.0003852922
#>  [8,] 7.992697e-05 2.172101e-05 7.649909e-05 7.548428e-05 0.0008840727
#>  [9,] 3.533981e-05 2.640122e-05 3.306342e-05 3.491242e-05 0.0004623678
#> [10,] 1.723592e-05 2.736532e-05 1.695174e-05 1.809313e-05 0.0003208529
#>                V60           V7           V8           V9
#>  [1,] 3.216859e-05 0.0005257964 0.0005062979 0.0005930592
#>  [2,] 1.483305e-05 0.0004485986 0.0003334365 0.0001430312
#>  [3,] 3.068728e-05 0.0003165844 0.0003844168 0.0006166601
#>  [4,] 3.981184e-05 0.0004416946 0.0005522130 0.0008974611
#>  [5,] 4.786962e-05 0.0005380469 0.0006426308 0.0010521709
#>  [6,] 5.821229e-06 0.0001901519 0.0001252306 0.0000644867
#>  [7,] 2.115557e-05 0.0004772895 0.0003909136 0.0003113311
#>  [8,] 7.481964e-05 0.0006428773 0.0009537906 0.0017623653
#>  [9,] 3.073239e-05 0.0004896945 0.0004727735 0.0005349975
#> [10,] 1.355151e-05 0.0004268532 0.0003197668 0.0001308486
#> 
#> $vW[[2]]
#>              [,1]        [,2]       [,3]         [,4]         [,5]         [,6]
#> [1,] -0.007074752 -0.01576881 -0.0178637  0.002321575 -0.008583518  0.001508755
#> [2,]  0.007074752  0.01576881  0.0178637 -0.002321575  0.008583518 -0.001508755
#>             [,7]       [,8]         [,9]       [,10]
#> [1,] -0.01323257  0.0131925 -0.009832106 -0.01705008
#> [2,]  0.01323257 -0.0131925  0.009832106  0.01705008
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.4933862 0.3836449 0.8300665 0.3334912 0.4469105 0.1609933 0.3166792
#>  [8] 0.5006164 0.6742800 0.3322951
#> 
#> $B[[2]]
#> [1] 0.003374428 0.027001530
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.004373600 0.003741050 0.002725734 0.003518151 0.004483552 0.001743722
#>  [7] 0.003981761 0.005155651 0.004068734 0.003551339
#> 
#> $vB[[2]]
#> [1] -0.02151003  0.02151003
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#>  [2,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>  [3,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#>  [4,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [5,] 0.0968 0.5025 0.3488 0.1700 0.2076 0.3087 0.4224 0.5312 0.2436 0.1884
#>  [6,] 0.0225 0.1452 0.1442 0.0948 0.0618 0.1641 0.0708 0.0844 0.2590 0.2679
#>  [7,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#>  [8,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#>  [9,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [10,] 0.0050 0.2282 0.2521 0.3484 0.3309 0.2614 0.1782 0.2055 0.2298 0.3545
#> [11,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [12,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [13,] 0.0253 0.2197 0.2653 0.3223 0.5582 0.6916 0.7943 0.7152 0.3512 0.2008
#> [14,] 0.0856 0.2751 0.2707 0.0946 0.1020 0.4519 0.6737 0.6699 0.7066 0.5632
#> [15,] 0.0310 0.4432 0.5222 0.5611 0.5379 0.4048 0.2245 0.1784 0.2297 0.2720
#> [16,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [17,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [18,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [19,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [20,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#> [21,] 0.0100 0.2990 0.3242 0.3565 0.3951 0.5201 0.6953 0.8468 1.0000 0.9278
#> [22,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [23,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [24,] 0.0307 0.2792 0.3877 0.4992 0.4981 0.4972 0.5607 0.7339 0.8230 0.9173
#> [25,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [26,] 0.0201 0.1199 0.1742 0.1387 0.2042 0.2580 0.2616 0.2097 0.2532 0.3213
#> [27,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [28,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#> [29,] 0.0491 0.2028 0.0947 0.2497 0.2209 0.3195 0.3340 0.3323 0.2780 0.2975
#> [30,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [31,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [32,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [33,] 0.0211 0.0568 0.0869 0.1935 0.1478 0.1871 0.1994 0.3283 0.6861 0.5814
#> [34,] 0.0240 0.0389 0.1934 0.2434 0.2906 0.2606 0.3811 0.4997 0.3015 0.3655
#> [35,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#> [36,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [37,] 0.0260 0.1630 0.1030 0.2187 0.1542 0.2630 0.2940 0.2978 0.0699 0.1401
#> [38,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [39,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [40,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#>  [2,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>  [3,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#>  [4,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [5,] 0.1908 0.0821 0.8321 1.0000 0.4076 0.0960 0.1928 0.2419 0.3790 0.2893
#>  [6,] 0.3094 0.0019 0.4678 0.5958 0.7245 0.8773 0.9214 0.9282 0.9942 1.0000
#>  [7,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#>  [8,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#>  [9,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [10,] 0.6218 0.0017 0.7265 0.8346 0.8268 0.8366 0.9408 0.9510 0.9801 0.9974
#> [11,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [12,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [13,] 0.2676 0.0808 0.4299 0.5280 0.3489 0.1430 0.5453 0.6338 0.7712 0.6838
#> [14,] 0.3785 0.0454 0.2721 0.5297 0.7697 0.8643 0.9304 0.9372 0.6247 0.6024
#> [15,] 0.5209 0.0221 0.6898 0.8202 0.8780 0.7600 0.7616 0.7152 0.7288 0.8686
#> [16,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [17,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [18,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [19,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [20,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#> [21,] 0.8510 0.0194 0.8010 0.8142 0.8825 0.7302 0.6107 0.7159 0.8458 0.6319
#> [22,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [23,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [24,] 0.9975 0.0523 0.9911 0.8240 0.6498 0.5980 0.4862 0.3150 0.1543 0.0989
#> [25,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [26,] 0.4327 0.0116 0.4760 0.5328 0.6057 0.6696 0.7476 0.8930 0.9405 1.0000
#> [27,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [28,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#> [29,] 0.2948 0.0279 0.1729 0.3264 0.3834 0.3523 0.5410 0.5228 0.4475 0.5340
#> [30,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [31,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [32,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [33,] 0.2500 0.0319 0.1734 0.3363 0.5588 0.6592 0.7012 0.8099 0.8901 0.8745
#> [34,] 0.6791 0.0218 0.7307 0.5053 0.4441 0.6987 0.8133 0.7781 0.8943 0.8929
#> [35,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#> [36,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [37,] 0.2990 0.0192 0.3915 0.3598 0.2403 0.4208 0.5675 0.6094 0.6323 0.6549
#> [38,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [39,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [40,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#>  [2,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>  [3,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#>  [4,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [5,] 0.3451 0.3777 0.0629 0.5213 0.2316 0.3335 0.4781 0.6116 0.6705 0.7375
#>  [6,] 0.9071 0.8545 0.0075 0.7293 0.6499 0.6071 0.5588 0.5967 0.6275 0.5459
#>  [7,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#>  [8,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#>  [9,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [10,] 1.0000 0.9036 0.0270 0.6409 0.3857 0.2908 0.2040 0.1653 0.1769 0.1140
#> [11,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [12,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [13,] 0.8015 0.8073 0.0507 0.8310 0.7792 0.5049 0.1413 0.2767 0.5084 0.4787
#> [14,] 0.6810 0.5047 0.0382 0.5775 0.4754 0.2400 0.2779 0.1997 0.5305 0.7409
#> [15,] 0.9509 0.8348 0.0433 0.5730 0.4363 0.4289 0.4240 0.3156 0.1287 0.1477
#> [16,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [17,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [18,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [19,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [20,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#> [21,] 0.4808 0.6291 0.0155 0.7152 0.6005 0.4235 0.4106 0.3992 0.1730 0.1975
#> [22,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [23,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [24,] 0.0284 0.1008 0.0653 0.2636 0.2694 0.2930 0.2925 0.3998 0.3660 0.3172
#> [25,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [26,] 0.9785 0.8473 0.0123 0.7639 0.6701 0.4989 0.3718 0.2196 0.1416 0.2680
#> [27,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [28,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#> [29,] 0.5323 0.3907 0.0592 0.3456 0.4091 0.4639 0.5580 0.5727 0.6355 0.7563
#> [30,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [31,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [32,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [33,] 0.7887 0.8725 0.0415 0.9376 0.8920 0.7508 0.6832 0.7610 0.9017 1.0000
#> [34,] 0.8913 0.8610 0.0324 0.8063 0.5540 0.2446 0.3459 0.1615 0.2467 0.5564
#> [35,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#> [36,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [37,] 0.7673 1.0000 0.0254 0.8463 0.5509 0.4444 0.5169 0.4268 0.1802 0.0791
#> [38,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [39,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [40,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#>  [2,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>  [3,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#>  [4,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [5,] 0.7356 0.7792 0.6788 0.0608 0.5259 0.2762 0.1545 0.2019 0.2231 0.4221
#>  [6,] 0.4786 0.3965 0.2087 0.0097 0.1651 0.1836 0.0652 0.0758 0.0486 0.0353
#>  [7,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#>  [8,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#>  [9,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [10,] 0.0740 0.0941 0.0621 0.0450 0.0426 0.0572 0.1068 0.1909 0.2229 0.2203
#> [11,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [12,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [13,] 0.1356 0.2299 0.2789 0.0244 0.3833 0.2933 0.1155 0.1705 0.1294 0.0909
#> [14,] 0.7775 0.4424 0.1416 0.0203 0.3508 0.4482 0.4208 0.3054 0.2235 0.2611
#> [15,] 0.2062 0.2400 0.5173 0.0191 0.5168 0.1491 0.2407 0.3415 0.4494 0.4624
#> [16,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [17,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [18,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [19,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [20,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#> [21,] 0.2370 0.1339 0.1583 0.0489 0.3151 0.1968 0.2054 0.1272 0.1129 0.1946
#> [22,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [23,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [24,] 0.4609 0.4374 0.1820 0.0521 0.3376 0.6202 0.4448 0.1863 0.1420 0.0589
#> [25,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [26,] 0.2630 0.3104 0.3392 0.0245 0.2123 0.1170 0.2655 0.2203 0.1541 0.1464
#> [27,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [28,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#> [29,] 0.6903 0.6176 0.5379 0.1270 0.5622 0.6508 0.4797 0.3736 0.2804 0.1982
#> [30,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [31,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [32,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [33,] 0.9123 0.7388 0.5915 0.0286 0.4057 0.3019 0.2331 0.2931 0.2298 0.2391
#> [34,] 0.4681 0.0979 0.1582 0.0569 0.0751 0.3321 0.3745 0.2666 0.1078 0.1418
#> [35,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#> [36,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [37,] 0.0535 0.1906 0.2561 0.0061 0.2153 0.2769 0.2841 0.1733 0.0815 0.0335
#> [38,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [39,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [40,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#>  [2,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>  [3,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#>  [4,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [5,] 0.3067 0.1329 0.1349 0.1057 0.0617 0.0499 0.0206 0.0073 0.0081 0.0303
#>  [6,] 0.0297 0.0241 0.0379 0.0119 0.0445 0.0073 0.0051 0.0034 0.0129 0.0100
#>  [7,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#>  [8,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#>  [9,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [10,] 0.2265 0.1766 0.1097 0.0558 0.0958 0.0142 0.0281 0.0165 0.0056 0.0010
#> [11,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [12,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [13,] 0.0800 0.0567 0.0198 0.0114 0.1724 0.0151 0.0085 0.0178 0.0073 0.0079
#> [14,] 0.2798 0.2392 0.2021 0.1326 0.0385 0.0358 0.0128 0.0172 0.0138 0.0079
#> [15,] 0.2001 0.0775 0.1232 0.0783 0.0964 0.0089 0.0249 0.0204 0.0059 0.0053
#> [16,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [17,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [18,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [19,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [20,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#> [21,] 0.2195 0.1930 0.1498 0.0773 0.0839 0.0196 0.0122 0.0130 0.0073 0.0077
#> [22,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [23,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [24,] 0.0576 0.0672 0.0269 0.0245 0.0611 0.0190 0.0063 0.0321 0.0189 0.0137
#> [25,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [26,] 0.1044 0.1225 0.0745 0.0490 0.0547 0.0224 0.0032 0.0076 0.0045 0.0056
#> [27,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [28,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#> [29,] 0.2438 0.1789 0.1706 0.0762 0.1772 0.0238 0.0268 0.0081 0.0129 0.0161
#> [30,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [31,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [32,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [33,] 0.1910 0.1096 0.0300 0.0171 0.0121 0.0383 0.0053 0.0090 0.0042 0.0153
#> [34,] 0.1687 0.0738 0.0634 0.0144 0.0330 0.0226 0.0061 0.0162 0.0146 0.0093
#> [35,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#> [36,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [37,] 0.0933 0.1018 0.0309 0.0208 0.0352 0.0318 0.0132 0.0118 0.0120 0.0051
#> [38,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [39,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [40,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#>  [2,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#>  [3,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#>  [4,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [5,] 0.0190 0.0212 0.0126 0.0201 0.0210 0.1207 0.0041 0.0944 0.4223 0.5744
#>  [6,] 0.0044 0.0057 0.0030 0.0035 0.0021 0.0906 0.0027 0.0889 0.0655 0.1624
#>  [7,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#>  [8,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#>  [9,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [10,] 0.0027 0.0062 0.0024 0.0063 0.0017 0.0830 0.0028 0.0879 0.1220 0.1977
#> [11,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [12,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [13,] 0.0038 0.0116 0.0033 0.0039 0.0081 0.3823 0.0053 0.3729 0.3583 0.3429
#> [14,] 0.0037 0.0051 0.0258 0.0102 0.0037 0.0534 0.0037 0.2140 0.3110 0.2837
#> [15,] 0.0079 0.0037 0.0015 0.0056 0.0067 0.1827 0.0054 0.1106 0.1702 0.2804
#> [16,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [17,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [18,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [19,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [20,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#> [21,] 0.0075 0.0060 0.0080 0.0019 0.0053 0.1009 0.0019 0.1627 0.2071 0.2696
#> [22,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [23,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [24,] 0.0277 0.0152 0.0052 0.0121 0.0124 0.0577 0.0055 0.0665 0.0664 0.1460
#> [25,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [26,] 0.0075 0.0037 0.0045 0.0029 0.0008 0.0208 0.0018 0.0891 0.0836 0.1335
#> [27,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [28,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#> [29,] 0.0063 0.0119 0.0194 0.0140 0.0332 0.1908 0.0439 0.2217 0.0768 0.1246
#> [30,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [31,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [32,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [33,] 0.0106 0.0020 0.0105 0.0049 0.0070 0.0438 0.0080 0.1299 0.1390 0.0695
#> [34,] 0.0112 0.0094 0.0054 0.0019 0.0066 0.0513 0.0023 0.0897 0.0713 0.0569
#> [35,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#> [36,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [37,] 0.0070 0.0015 0.0035 0.0008 0.0044 0.0701 0.0077 0.1263 0.1080 0.1523
#> [38,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [39,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [40,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6658039 0.7768602 0.9211614 0.4012500 0.6343680 0.3032864 0.7104053
#>  [2,] 0.6226583 0.5625535 0.8739322 0.3980947 0.4489461 0.1758038 0.5212726
#>  [3,] 0.6434777 0.7182121 0.8723080 0.5391608 0.6537018 0.4124696 0.6721748
#>  [4,] 0.7319034 0.7456852 0.9258002 0.5045655 0.6433998 0.2861388 0.6420366
#>  [5,] 0.5811493 0.4877645 0.7776246 0.3117591 0.3643758 0.1545562 0.4183572
#>  [6,] 0.7430485 0.8018019 0.9430462 0.5320015 0.6965027 0.3191530 0.6853358
#>  [7,] 0.6098103 0.6959872 0.8768216 0.3451907 0.5481720 0.2341195 0.6309809
#>  [8,] 0.5821899 0.5735453 0.8657466 0.3205420 0.4149697 0.1368039 0.4953597
#>  [9,] 0.6574283 0.7593676 0.9178454 0.4163188 0.6346579 0.3011346 0.7022077
#> [10,] 0.6891546 0.8172628 0.9252565 0.4746403 0.7135347 0.3431690 0.7217880
#> [11,] 0.6538453 0.6918877 0.8795588 0.4362667 0.5704015 0.3045506 0.6250798
#> [12,] 0.6703270 0.8084041 0.9303739 0.4010888 0.6750200 0.3078733 0.7341994
#> [13,] 0.5911520 0.6310376 0.8552409 0.3649793 0.5222196 0.2055978 0.5685861
#> [14,] 0.5991069 0.6625947 0.8694491 0.3342163 0.5061979 0.1922362 0.5911820
#> [15,] 0.6401463 0.7319298 0.8976511 0.3849722 0.5727296 0.2286821 0.6291492
#> [16,] 0.6280768 0.7535014 0.9026190 0.3912448 0.6353665 0.2883748 0.6908604
#> [17,] 0.6440408 0.7766929 0.9095712 0.3775116 0.6432167 0.3064006 0.7135092
#> [18,] 0.6853100 0.6459785 0.8686198 0.4979393 0.5630229 0.2889930 0.5601125
#> [19,] 0.7011258 0.7968740 0.9184210 0.5079430 0.6977746 0.3878005 0.7191801
#> [20,] 0.6529195 0.7310567 0.8977633 0.3625036 0.5933552 0.2944082 0.6677038
#> [21,] 0.6351393 0.7221646 0.9056494 0.3418796 0.5819755 0.2490308 0.6713400
#> [22,] 0.6946762 0.7703665 0.9153019 0.4946560 0.6643157 0.3414447 0.6929627
#> [23,] 0.6512967 0.7064168 0.8890643 0.4744905 0.6087793 0.2914786 0.6271631
#> [24,] 0.6007191 0.5901838 0.8356459 0.3186700 0.4594686 0.2360901 0.5736245
#> [25,] 0.6930784 0.7279109 0.9117967 0.5001016 0.6256214 0.2821459 0.6229567
#> [26,] 0.7027574 0.7862422 0.9247295 0.4931348 0.6755637 0.3326804 0.7031039
#> [27,] 0.6994509 0.7884125 0.9229566 0.5178568 0.6940498 0.3472882 0.7077378
#> [28,] 0.5445099 0.6010418 0.7911929 0.2817491 0.4827761 0.2246877 0.5442170
#> [29,] 0.5665798 0.5159925 0.8091647 0.3465768 0.3946235 0.1671378 0.4712109
#> [30,] 0.6092531 0.6944913 0.8702384 0.3561052 0.5653239 0.2648453 0.6423285
#> [31,] 0.6591319 0.7439275 0.8983332 0.4818700 0.6408582 0.3183311 0.6651399
#> [32,] 0.6630435 0.7906613 0.9214622 0.4101834 0.6602514 0.3126062 0.7176627
#> [33,] 0.7056485 0.6956293 0.9257438 0.4154771 0.5455592 0.1845887 0.5840563
#> [34,] 0.6816061 0.7712689 0.9172655 0.4556936 0.6617652 0.3084912 0.6920727
#> [35,] 0.6800355 0.7111731 0.8926852 0.4374175 0.5767557 0.2436300 0.6099277
#> [36,] 0.6401761 0.6979954 0.8929631 0.3556880 0.5180526 0.2033732 0.5978309
#> [37,] 0.6679404 0.7068984 0.8931148 0.5055278 0.6231624 0.3330353 0.6344592
#> [38,] 0.6947972 0.8063062 0.9289453 0.4301087 0.6962599 0.3802245 0.7458256
#> [39,] 0.6446182 0.7502316 0.8983487 0.3874572 0.5996481 0.2600285 0.6490122
#> [40,] 0.6235923 0.7138628 0.8822356 0.4447981 0.6156840 0.3230057 0.6634134
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4494826 0.7940893 0.8141277
#>  [2,] 0.5399078 0.6122603 0.6858360
#>  [3,] 0.5912626 0.7197381 0.7762672
#>  [4,] 0.6329217 0.7425535 0.8126750
#>  [5,] 0.4119233 0.5636475 0.5846161
#>  [6,] 0.6404009 0.7861259 0.8518847
#>  [7,] 0.4652040 0.7443958 0.7361624
#>  [8,] 0.4377419 0.5633512 0.6996568
#>  [9,] 0.4567567 0.7860923 0.8068697
#> [10,] 0.4917110 0.7817520 0.8562140
#> [11,] 0.4950768 0.7104685 0.7515952
#> [12,] 0.4213279 0.8066785 0.8478832
#> [13,] 0.4692809 0.6923009 0.7351720
#> [14,] 0.4559853 0.7093263 0.7194075
#> [15,] 0.4088824 0.6803842 0.7966588
#> [16,] 0.4072270 0.7705844 0.8048802
#> [17,] 0.3932583 0.7993001 0.8114795
#> [18,] 0.6268799 0.6780677 0.7229491
#> [19,] 0.5664465 0.7836653 0.8378607
#> [20,] 0.3938319 0.7916852 0.7586992
#> [21,] 0.3822036 0.7721905 0.7772699
#> [22,] 0.5647857 0.7478394 0.8291278
#> [23,] 0.5320160 0.6720584 0.7884619
#> [24,] 0.3479891 0.7357923 0.6266188
#> [25,] 0.6221386 0.7161280 0.8013290
#> [26,] 0.5666138 0.7563345 0.8456831
#> [27,] 0.5674464 0.7679366 0.8406413
#> [28,] 0.3245747 0.7159271 0.6223576
#> [29,] 0.4769381 0.5800508 0.6170580
#> [30,] 0.4326895 0.7619363 0.7237994
#> [31,] 0.5419326 0.7158474 0.8128023
#> [32,] 0.4446330 0.7962389 0.8314400
#> [33,] 0.6282404 0.7184353 0.7864009
#> [34,] 0.5435446 0.7781731 0.8254165
#> [35,] 0.5748875 0.6920672 0.7824785
#> [36,] 0.4146310 0.6455794 0.7836866
#> [37,] 0.5772230 0.7082176 0.7907627
#> [38,] 0.4565473 0.8251264 0.8310471
#> [39,] 0.4144747 0.7340680 0.7890763
#> [40,] 0.4495945 0.7356992 0.7793120
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.3928370 0.6071630
#>  [2,] 0.3931269 0.6068731
#>  [3,] 0.3823769 0.6176231
#>  [4,] 0.3781952 0.6218048
#>  [5,] 0.4081671 0.5918329
#>  [6,] 0.3753617 0.6246383
#>  [7,] 0.3967952 0.6032048
#>  [8,] 0.4037320 0.5962680
#>  [9,] 0.3920173 0.6079827
#> [10,] 0.3867458 0.6132542
#> [11,] 0.3920583 0.6079417
#> [12,] 0.3935010 0.6064990
#> [13,] 0.3975172 0.6024828
#> [14,] 0.3989565 0.6010435
#> [15,] 0.3989472 0.6010528
#> [16,] 0.3965324 0.6034676
#> [17,] 0.3971522 0.6028478
#> [18,] 0.3823514 0.6176486
#> [19,] 0.3817754 0.6182246
#> [20,] 0.3983160 0.6016840
#> [21,] 0.3999940 0.6000060
#> [22,] 0.3832480 0.6167520
#> [23,] 0.3883081 0.6116919
#> [24,] 0.4065968 0.5934032
#> [25,] 0.3803389 0.6196611
#> [26,] 0.3826390 0.6173610
#> [27,] 0.3813666 0.6186334
#> [28,] 0.4107864 0.5892136
#> [29,] 0.4019554 0.5980446
#> [30,] 0.3979516 0.6020484
#> [31,] 0.3864269 0.6135731
#> [32,] 0.3924715 0.6075285
#> [33,] 0.3830775 0.6169225
#> [34,] 0.3851970 0.6148030
#> [35,] 0.3868464 0.6131536
#> [36,] 0.4010077 0.5989923
#> [37,] 0.3838239 0.6161761
#> [38,] 0.3896719 0.6103281
#> [39,] 0.3972920 0.6027080
#> [40,] 0.3937438 0.6062562
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]        [,2]     [,3]         [,4]        [,5]       [,6]
#>  [1,] 0.6892670  1.24746182 2.458232 -0.400259361  0.55100189 -0.8316968
#>  [2,] 0.5008466  0.25153175 1.936183 -0.413410187 -0.20492984 -1.5450402
#>  [3,] 0.5904905  0.93561004 1.921522  0.156964550  0.63535091 -0.3537655
#>  [4,] 1.0043011  1.07573046 2.523897  0.018262472  0.59015087 -0.9142113
#>  [5,] 0.3274933 -0.04895188 1.251877 -0.791908168 -0.55642167 -1.6993037
#>  [6,] 1.0618740  1.39759446 2.806875  0.128181351  0.83069892 -0.7576668
#>  [7,] 0.4465147  0.82826135 1.962669 -0.640246974  0.19328765 -1.1851947
#>  [8,] 0.3317696  0.29633096 1.863863 -0.751281938 -0.34345829 -1.8420935
#>  [9,] 0.6518547  1.14921564 2.413427 -0.337903775  0.55225189 -0.8419009
#> [10,] 0.7961698  1.49791197 2.516009 -0.101525833  0.91261354 -0.6492037
#> [11,] 0.6359850  0.80895948 1.988259 -0.256327553  0.28348925 -0.8257211
#> [12,] 0.7096642  1.43967370 2.592448 -0.400930408  0.73097875 -0.8100805
#> [13,] 0.3687296  0.53667078 1.776312 -0.553816940  0.08893702 -1.3516683
#> [14,] 0.4017454  0.67487848 1.896097 -0.689176251  0.02479306 -1.4355447
#> [15,] 0.5759993  1.00443584 2.171393 -0.468496175  0.29299670 -1.2157678
#> [16,] 0.5239751  1.11737437 2.226669 -0.442082606  0.55530909 -0.9032905
#> [17,] 0.5929461  1.24649731 2.308410 -0.500123531  0.58935328 -0.8170009
#> [18,] 0.7782834  0.60140843 1.888810 -0.008242797  0.25343962 -0.9002798
#> [19,] 0.8526647  1.36686995 2.421084  0.031774826  0.83672294 -0.4565672
#> [20,] 0.6318969  0.99999067 2.172616 -0.564514438  0.37785331 -0.8740697
#> [21,] 0.5543288  0.95522426 2.261635 -0.654929458  0.33088844 -1.1037880
#> [22,] 0.8220729  1.21038192 2.380160 -0.021376650  0.68258622 -0.6568630
#> [23,] 0.6247440  0.87804447 2.081219 -0.102126701  0.44218392 -0.8882136
#> [24,] 0.4084624  0.36472527 1.626182 -0.759890824 -0.16248208 -1.1742363
#> [25,] 0.8145506  0.98404928 2.335773  0.000406571  0.51347771 -0.9338418
#> [26,] 0.8604630  1.30242137 2.508413 -0.027462510  0.73345845 -0.6960868
#> [27,] 0.8446845  1.31538251 2.483213  0.071457724  0.81912123 -0.6309809
#> [28,] 0.1785121  0.40980785 1.332131 -0.935801827 -0.06892271 -1.2385544
#> [29,] 0.2679102  0.06399195 1.444591 -0.634120654 -0.42791862 -1.6060498
#> [30,] 0.4441735  0.82120143 1.903068 -0.592309195  0.26279782 -1.0209349
#> [31,] 0.6594280  1.06648317 2.178840 -0.072551644  0.57909087 -0.7614522
#> [32,] 0.6768866  1.32891624 2.462382 -0.363207286  0.66441484 -0.7879632
#> [33,] 0.8743426  0.82657031 2.523077 -0.341368269  0.18274384 -1.4855627
#> [34,] 0.7611627  1.21548992 2.405760 -0.177691548  0.67117057 -0.8071826
#> [35,] 0.7539348  0.90108841 2.118468 -0.251649854  0.30946919 -1.1328801
#> [36,] 0.5761284  0.83777021 2.121371 -0.594129129  0.07224188 -1.3653434
#> [37,] 0.6988843  0.88036746 2.122960  0.022112275  0.50299283 -0.6944888
#> [38,] 0.8226436  1.42618506 2.570599 -0.281407606  0.82955082 -0.4885954
#> [39,] 0.5954656  1.09984789 2.179010 -0.458013219  0.40399917 -1.0458203
#> [40,] 0.5048239  0.91421925 2.013773 -0.221711273  0.47126887 -0.7399928
#>              [,7]        [,8]      [,9]     [,10]
#>  [1,]  0.89735351 -0.20276152 1.3497532 1.4770571
#>  [2,]  0.08514165  0.15997128 0.4568234 0.7807235
#>  [3,]  0.71803792  0.36918741 0.9431631 1.2440444
#>  [4,]  0.58421473  0.54477166 1.0592831 1.4674861
#>  [5,] -0.32952078 -0.35602002 0.2559787 0.3417523
#>  [6,]  0.77840306  0.57710448 1.3017293 1.7494599
#>  [7,]  0.53642722 -0.13940928 1.0689429 1.0261167
#>  [8,] -0.01856178 -0.25033167 0.2547741 0.8456642
#>  [9,]  0.85783292 -0.17340637 1.3015297 1.4297974
#> [10,]  0.95334825 -0.03315884 1.2759053 1.7841946
#> [11,]  0.51116591 -0.01969327 0.8976607 1.1071380
#> [12,]  1.01603402 -0.31732465 1.4285708 1.7180941
#> [13,]  0.27608497 -0.12303152 0.8108984 1.0210242
#> [14,]  0.36885388 -0.17651564 0.8921145 0.9415247
#> [15,]  0.52856866 -0.36858730 0.7555382 1.3655413
#> [16,]  0.80414478 -0.37544065 1.2116142 1.4170800
#> [17,]  0.91248914 -0.43363658 1.3819254 1.4596527
#> [18,]  0.24161875  0.51885455 0.7449060 0.9591375
#> [19,]  0.94039839  0.26736759 1.2871555 1.6423960
#> [20,]  0.69781776 -0.43123307 1.3351137 1.1455613
#> [21,]  0.71425191 -0.48020534 1.2207213 1.2498271
#> [22,]  0.81400709  0.26060770 1.0871221 1.5794585
#> [23,]  0.52006568  0.12823927 0.7175096 1.3156792
#> [24,]  0.29665442 -0.62788998 1.0242122 0.5177384
#> [25,]  0.50211706  0.49863548 0.9253357 1.3946211
#> [26,]  0.86212240  0.26804861 1.1326871 1.7011367
#> [27,]  0.88442253  0.27143993 1.1966968 1.6630073
#> [28,]  0.17733122 -0.73282679 0.9243473 0.4995671
#> [29,] -0.11528394 -0.09231324 0.3229818 0.4770793
#> [30,]  0.58548467 -0.27088652 1.1633248 0.9633871
#> [31,]  0.68628464  0.16812527 0.9239555 1.4683227
#> [32,]  0.93289721 -0.22237987 1.3629507 1.5958675
#> [33,]  0.33944727  0.52467541 0.9367135 1.3033657
#> [34,]  0.80982734  0.17462089 1.2550518 1.5534851
#> [35,]  0.44700831  0.30182050 0.8098013 1.2801688
#> [36,]  0.39643522 -0.34485334 0.5996639 1.2872810
#> [37,]  0.55139516  0.31138386 0.8867429 1.3295291
#> [38,]  1.07647118 -0.17425044 1.5514733 1.5930664
#> [39,]  0.61470021 -0.34549727 1.0153611 1.3193668
#> [40,]  0.67854264 -0.20230915 1.0237334 1.2616618
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.2097601 0.2256423
#>  [2,] -0.2001825 0.2340047
#>  [3,] -0.2257503 0.2537215
#>  [4,] -0.2302547 0.2669610
#>  [5,] -0.1724101 0.1991376
#>  [6,] -0.2371752 0.2721074
#>  [7,] -0.1975632 0.2212732
#>  [8,] -0.1805648 0.2093740
#>  [9,] -0.2108899 0.2279508
#> [10,] -0.2198978 0.2411139
#> [11,] -0.2083016 0.2303669
#> [12,] -0.2092892 0.2233303
#> [13,] -0.1938875 0.2219334
#> [14,] -0.1924593 0.2173556
#> [15,] -0.1956466 0.2142071
#> [16,] -0.2023165 0.2176182
#> [17,] -0.2027915 0.2145535
#> [18,] -0.2207807 0.2587988
#> [19,] -0.2285875 0.2534319
#> [20,] -0.2003289 0.2121578
#> [21,] -0.1962544 0.2092356
#> [22,] -0.2244034 0.2513811
#> [23,] -0.2127093 0.2417205
#> [24,] -0.1846061 0.1934462
#> [25,] -0.2254122 0.2626977
#> [26,] -0.2255033 0.2528586
#> [27,] -0.2285384 0.2552135
#> [28,] -0.1731403 0.1875750
#> [29,] -0.1832932 0.2140310
#> [30,] -0.1976048 0.2164024
#> [31,] -0.2174295 0.2449271
#> [32,] -0.2103090 0.2266262
#> [33,] -0.2177733 0.2587330
#> [34,] -0.2202480 0.2472991
#> [35,] -0.2131538 0.2474339
#> [36,] -0.1906316 0.2106363
#> [37,] -0.2215143 0.2518345
#> [38,] -0.2172491 0.2314425
#> [39,] -0.1997082 0.2170530
#> [40,] -0.2071432 0.2244591
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.6071630 -0.6071630
#>  [2,] -0.3931269  0.3931269
#>  [3,] -0.3823769  0.3823769
#>  [4,] -0.3781952  0.3781952
#>  [5,]  0.5918329 -0.5918329
#>  [6,] -0.3753617  0.3753617
#>  [7,] -0.3967952  0.3967952
#>  [8,]  0.5962680 -0.5962680
#>  [9,]  0.6079827 -0.6079827
#> [10,]  0.6132542 -0.6132542
#> [11,] -0.3920583  0.3920583
#> [12,]  0.6064990 -0.6064990
#> [13,] -0.3975172  0.3975172
#> [14,] -0.3989565  0.3989565
#> [15,]  0.6010528 -0.6010528
#> [16,]  0.6034676 -0.6034676
#> [17,]  0.6028478 -0.6028478
#> [18,] -0.3823514  0.3823514
#> [19,] -0.3817754  0.3817754
#> [20,]  0.6016840 -0.6016840
#> [21,] -0.3999940  0.3999940
#> [22,] -0.3832480  0.3832480
#> [23,]  0.6116919 -0.6116919
#> [24,]  0.5934032 -0.5934032
#> [25,] -0.3803389  0.3803389
#> [26,] -0.3826390  0.3826390
#> [27,]  0.6186334 -0.6186334
#> [28,] -0.4107864  0.4107864
#> [29,]  0.5980446 -0.5980446
#> [30,] -0.3979516  0.3979516
#> [31,]  0.6135731 -0.6135731
#> [32,]  0.6075285 -0.6075285
#> [33,] -0.3830775  0.3830775
#> [34,] -0.3851970  0.3851970
#> [35,]  0.6131536 -0.6131536
#> [36,]  0.5989923 -0.5989923
#> [37,] -0.3838239  0.3838239
#> [38,] -0.3896719  0.3896719
#> [39,]  0.6027080 -0.6027080
#> [40,]  0.6062562 -0.6062562
#> 
#> $L
#> [1] 0.7139388 0.6669677 0.8889832 0.6926164 0.7717087 0.7104501
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174