Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1          V10           V11          V12          V13
#>  [1,] -0.2898498 -0.051904504 -0.0363936064  0.001365970 -0.009887545
#>  [2,] -0.3988639 -0.042989086 -0.0559209106 -0.062314113 -0.018642327
#>  [3,] -0.2379521 -0.212504811 -0.1384642522 -0.082299082 -0.061831880
#>  [4,] -0.2615206 -0.077436953 -0.1743788484 -0.156092593 -0.060102781
#>  [5,] -0.2834501  0.001204718  0.0002308417 -0.065651127 -0.048962432
#>  [6,] -0.2576129 -0.132096500 -0.0449044134 -0.054021806  0.018802515
#>  [7,] -0.3539572 -0.097663422 -0.0838364130 -0.007811052 -0.106003418
#>  [8,] -0.2853015 -0.123367228 -0.0805894020 -0.066916516 -0.050959716
#>  [9,] -0.2263751 -0.129784469 -0.0638101842 -0.056137182 -0.122958226
#> [10,] -0.2590712 -0.225278815 -0.1692900800 -0.104325167 -0.166515228
#>                V14          V15         V16          V17         V18
#>  [1,] -0.128759938 -0.062368227 -0.08851350 -0.002682195  0.09240216
#>  [2,] -0.103786774  0.073278169  0.09891815 -0.015189193  0.10327643
#>  [3,] -0.004062889 -0.055086538 -0.07861051  0.100984882 -0.08595671
#>  [4,] -0.081064232  0.037992731  0.13171292  0.045234967  0.15034855
#>  [5,]  0.007937267 -0.152405348  0.01928571 -0.038125554 -0.01008171
#>  [6,] -0.006937515 -0.089523722  0.01768080 -0.049081816  0.07998088
#>  [7,]  0.056750222  0.036523367 -0.01335682  0.162357797  0.14007949
#>  [8,] -0.127582553  0.025508658  0.06151545 -0.016003384 -0.04249151
#>  [9,] -0.103436671  0.004303594  0.05656478  0.012380366  0.07243854
#> [10,] -0.026290584 -0.107201151  0.07518724  0.085749188  0.02916253
#>               V19         V2         V20          V21          V22
#>  [1,]  0.07810432 -0.2034072  0.05227105  0.150009363  0.064377026
#>  [2,] -0.03570145 -0.3869309  0.02944742  0.185235731  0.120467314
#>  [3,]  0.12415333 -0.2730016  0.09570431  0.150510672 -0.008259079
#>  [4,]  0.05333762 -0.2832136  0.01181250  0.075855561  0.144650716
#>  [5,] -0.02807262 -0.3476267  0.13831056  0.078538784  0.046962793
#>  [6,]  0.02133073 -0.2330309  0.01540926  0.055157691  0.192698778
#>  [7,]  0.10168620 -0.2071430  0.03516171  0.001784244 -0.028567158
#>  [8,]  0.05696498 -0.3492488  0.03849101  0.141908438  0.196205328
#>  [9,]  0.16730002 -0.1851996  0.15686778  0.164647792  0.057492352
#> [10,]  0.02147009 -0.3679027 -0.05673176 -0.088472735  0.031005259
#>                 V23        V24        V25        V26          V27        V28
#>  [1,]  0.1406231915 0.09622370 0.16709814 0.07765015  0.102140826 0.08616227
#>  [2,]  0.2007609076 0.11180885 0.14788497 0.17303107  0.134967760 0.14682035
#>  [3,]  0.0008564714 0.10979614 0.08266503 0.09630238  0.200862783 0.11303034
#>  [4,]  0.0685439105 0.02320245 0.10627921 0.05510991 -0.002637473 0.10401613
#>  [5,]  0.0705908942 0.14038960 0.06743179 0.21829292  0.223314935 0.11803110
#>  [6,]  0.1978158110 0.07440866 0.08090702 0.09381021  0.049605810 0.20711320
#>  [7,] -0.0495845202 0.01434283 0.01997782 0.11709228  0.133021224 0.17898693
#>  [8,]  0.1744020622 0.19196990 0.19651901 0.14087428  0.284144734 0.18125746
#>  [9,]  0.0250310245 0.04082398 0.13307169 0.10384614  0.142727561 0.02651605
#> [10,]  0.1309767982 0.08883617 0.19587887 0.16258854  0.133515925 0.19362110
#>               V29         V3         V30         V31         V32          V33
#>  [1,]  0.10383595 -0.3264123  0.01420535 -0.02710115 -0.01591845 -0.072681121
#>  [2,]  0.08158008 -0.2682201  0.10120634  0.09861744  0.04485017  0.051728540
#>  [3,]  0.09764589 -0.3017885  0.11868089  0.08876882 -0.02226069 -0.032995701
#>  [4,]  0.11350690 -0.2431976  0.11358420  0.14133260  0.08926136  0.048902534
#>  [5,]  0.13465787 -0.3306122  0.12755809 -0.05160288 -0.04334798  0.009646813
#>  [6,] -0.03530086 -0.2279233 -0.01917116  0.08587472  0.08598989  0.021196589
#>  [7,]  0.11432014 -0.2028205  0.09299014  0.06624773  0.12574194 -0.005826123
#>  [8,]  0.19040912 -0.4219097  0.16950188  0.02171230  0.05607372  0.044946887
#>  [9,] -0.01592100 -0.1763238 -0.12728023 -0.11499352 -0.18206518 -0.129510937
#> [10,]  0.08165425 -0.2898304 -0.05397972 -0.02655939  0.03652988 -0.074713183
#>                 V34          V35          V36          V37         V38
#>  [1,]  0.0594266157  0.133501150  0.067482257  0.032231461  0.04305813
#>  [2,]  0.0177401277 -0.055456196  0.094956095 -0.014445657 -0.07095830
#>  [3,]  0.0457322093  0.012087737 -0.055063609 -0.051832685  0.02053330
#>  [4,] -0.0508806376 -0.033602677 -0.106478231 -0.135149513 -0.02305172
#>  [5,]  0.0004898063  0.059625991 -0.008029357  0.077044746  0.10609460
#>  [6,]  0.0332584225 -0.009670725 -0.156223702 -0.095131794 -0.08482101
#>  [7,] -0.1035520791 -0.116976636 -0.082593396 -0.154846142 -0.01939446
#>  [8,]  0.0751922072 -0.054268211  0.087981243  0.056223798 -0.09305506
#>  [9,] -0.1771955725 -0.118283396 -0.163598560 -0.045780101 -0.17999485
#> [10,] -0.1533397249 -0.125051510 -0.133317469 -0.006342763 -0.02680514
#>                V39         V4          V40          V41         V42         V43
#>  [1,]  0.032487302 -0.1937808  0.045748217 -0.090003674 -0.09838501 -0.02917000
#>  [2,] -0.049118685 -0.3268907 -0.098255435 -0.096237655 -0.05887684 -0.01070844
#>  [3,] -0.047850291 -0.2885798  0.005608929 -0.159508244 -0.10910487 -0.09288642
#>  [4,] -0.107105004 -0.1198018  0.009908846  0.007865226 -0.14978859 -0.06086880
#>  [5,]  0.006278254 -0.1864373  0.003663248 -0.018472561 -0.01158448 -0.17749025
#>  [6,] -0.161887904 -0.3096473 -0.148561670 -0.117598677 -0.09010060 -0.21229504
#>  [7,]  0.056721138 -0.2940800 -0.006880884 -0.081146779 -0.08098064 -0.11925798
#>  [8,] -0.034145141 -0.4042391 -0.019565641  0.045975170 -0.01627941 -0.04656062
#>  [9,] -0.008700128 -0.2589895 -0.078046853 -0.093811057 -0.11422652 -0.17545716
#> [10,] -0.100951089 -0.2265414 -0.097120199 -0.097238562 -0.09172076 -0.06300417
#>                V44         V45         V46        V47        V48        V49
#>  [1,] -0.111985300 -0.16685075 -0.03598432 -0.1075082 -0.1765518 -0.2805526
#>  [2,] -0.066602733 -0.05213710 -0.20894194 -0.2245285 -0.1431572 -0.2683663
#>  [3,] -0.104099106 -0.03424047 -0.11082282 -0.1184371 -0.2851379 -0.2148447
#>  [4,] -0.138161415 -0.03720076 -0.10001705 -0.2154724 -0.1527123 -0.2474773
#>  [5,] -0.117037551 -0.03344904 -0.15110530 -0.1954153 -0.2424942 -0.2931544
#>  [6,] -0.202829063 -0.10869459 -0.25210892 -0.2379397 -0.2058221 -0.3737085
#>  [7,] -0.001227033 -0.04596327 -0.08679615 -0.1106413 -0.2356703 -0.2295697
#>  [8,]  0.035254091 -0.16419563 -0.18275173 -0.1820891 -0.2285820 -0.2645262
#>  [9,] -0.200821838 -0.16839338 -0.26850974 -0.2044950 -0.2517190 -0.3531896
#> [10,] -0.119056384 -0.22975550 -0.08956161 -0.2273883 -0.1767417 -0.2597086
#>                V5        V50        V51        V52        V53        V54
#>  [1,] -0.09995065 -0.2415823 -0.3487046 -0.3143857 -0.2508240 -0.2372919
#>  [2,] -0.26097001 -0.3704683 -0.3075493 -0.3739498 -0.3418191 -0.3459025
#>  [3,] -0.14922528 -0.2866488 -0.3849726 -0.3956619 -0.3246132 -0.3259720
#>  [4,] -0.18463604 -0.3203258 -0.2318764 -0.3299715 -0.3268666 -0.2001971
#>  [5,] -0.20731813 -0.3803357 -0.3575718 -0.2390152 -0.2921054 -0.4265707
#>  [6,] -0.23827239 -0.3335656 -0.3631493 -0.2784646 -0.2720548 -0.3207804
#>  [7,] -0.22273173 -0.2511376 -0.3540792 -0.4098426 -0.3505728 -0.4090144
#>  [8,] -0.34860261 -0.4674386 -0.4287345 -0.3034775 -0.3888328 -0.3044261
#>  [9,] -0.29475455 -0.2650770 -0.3343542 -0.3622270 -0.3303504 -0.3496634
#> [10,] -0.28647198 -0.2339714 -0.2102546 -0.2586787 -0.2808027 -0.2504726
#>              V55        V56        V57        V58        V59          V6
#>  [1,] -0.3344666 -0.2302744 -0.3275672 -0.3546555 -0.3514687 -0.23343416
#>  [2,] -0.3619739 -0.2789106 -0.3024407 -0.4295476 -0.2501649 -0.12596199
#>  [3,] -0.3892629 -0.3435967 -0.2550052 -0.3144225 -0.2428528 -0.26410277
#>  [4,] -0.2044301 -0.2119689 -0.2932007 -0.3098036 -0.1884236 -0.17875597
#>  [5,] -0.4477488 -0.3790838 -0.4163657 -0.2687447 -0.4191757 -0.20442064
#>  [6,] -0.2972609 -0.3070924 -0.3062247 -0.2662690 -0.3050392 -0.25259740
#>  [7,] -0.2378134 -0.4258622 -0.3952837 -0.3779675 -0.3074684 -0.07649277
#>  [8,] -0.3293812 -0.4432665 -0.3520360 -0.4737540 -0.4680809 -0.22193192
#>  [9,] -0.2123063 -0.3198839 -0.3368620 -0.2882793 -0.3677320 -0.20835031
#> [10,] -0.2629425 -0.3830907 -0.3116104 -0.3171182 -0.3643631 -0.18772010
#>              V60         V7          V8          V9
#>  [1,] -0.2648225 -0.1822468 -0.04260295 -0.12257370
#>  [2,] -0.2846810 -0.1146730 -0.10840631 -0.04905764
#>  [3,] -0.2957569 -0.1163530 -0.15171631 -0.09393086
#>  [4,] -0.2390779 -0.1938759 -0.15492558 -0.12422661
#>  [5,] -0.4451090 -0.2522681 -0.20696068 -0.18550710
#>  [6,] -0.2980335 -0.1778573 -0.16082486 -0.13116911
#>  [7,] -0.2988704 -0.1222570 -0.16572265 -0.12010145
#>  [8,] -0.4234153 -0.1943676 -0.18199599 -0.06202547
#>  [9,] -0.3526390 -0.1394110 -0.05819477 -0.14605713
#> [10,] -0.3139858 -0.2720512 -0.23851745 -0.09660744
#> 
#> $W[[2]]
#>             [,1]       [,2]        [,3]       [,4]        [,5]       [,6]
#> [1,] -0.07480025 0.05869783  0.04607576 0.03171517  0.11063689 0.00842191
#> [2,] -0.02878645 0.04591186 -0.06017698 0.04040380 -0.04767776 0.06864679
#>             [,7]        [,8]        [,9]      [,10]
#> [1,]  0.10416845 -0.03221723 -0.10006899 0.03176106
#> [2,] -0.08556887 -0.07665009 -0.02923359 0.04868605
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  5.398429e-05  4.442032e-04  6.303085e-04  5.757159e-04  5.369086e-04
#>  [2,]  8.550446e-06  1.200255e-04  1.538122e-04  1.503276e-04  1.796415e-04
#>  [3,] -9.113735e-05 -5.888601e-04 -9.129878e-04 -8.246627e-04 -6.280361e-04
#>  [4,]  1.176884e-05  1.651669e-04  2.222644e-04  2.082689e-04  2.329390e-04
#>  [5,] -1.151909e-04 -7.111335e-04 -1.090927e-03 -9.493838e-04 -7.424978e-04
#>  [6,]  3.760760e-05  3.819681e-04  5.730819e-04  5.318100e-04  4.875650e-04
#>  [7,] -1.518780e-04 -1.073197e-03 -1.646447e-03 -1.491798e-03 -1.188262e-03
#>  [8,] -6.208110e-06  1.661428e-05  2.249503e-06  2.481376e-05  6.507698e-05
#>  [9,]  5.126988e-05  4.820162e-04  7.389849e-04  7.011378e-04  6.039330e-04
#> [10,] -8.050446e-07  9.351098e-05  1.145866e-04  9.574004e-05  1.350053e-04
#>                 V14           V15           V16          V17          V18
#>  [1,]  0.0004826663  4.059291e-04  4.719117e-04 3.996808e-04 0.0003453410
#>  [2,]  0.0002141891  2.468246e-04  3.168676e-04 2.959100e-04 0.0002794231
#>  [3,] -0.0003771515 -5.431244e-05  8.523203e-05 1.726546e-04 0.0002209766
#>  [4,]  0.0002599703  2.910094e-04  3.835934e-04 3.575056e-04 0.0003383905
#>  [5,] -0.0004821996 -1.461744e-04 -1.789951e-05 1.259178e-04 0.0002221124
#>  [6,]  0.0004264916  3.621364e-04  4.669171e-04 4.326203e-04 0.0004046695
#>  [7,] -0.0007938159 -2.958647e-04 -1.471547e-04 8.001569e-06 0.0001197549
#>  [8,]  0.0001134384  1.639263e-04  2.252180e-04 2.337246e-04 0.0002357502
#>  [9,]  0.0004873603  3.655774e-04  4.413428e-04 4.110891e-04 0.0004072618
#> [10,]  0.0001844860  2.541199e-04  3.770198e-04 3.615419e-04 0.0003479632
#>                 V19            V2           V20           V21           V22
#>  [1,]  0.0008028767  9.078714e-05  0.0010433605  0.0011587077  0.0011705994
#>  [2,]  0.0003395896  1.521755e-05  0.0003733685  0.0004089927  0.0004583989
#>  [3,] -0.0006272301 -1.464449e-04 -0.0010248231 -0.0011236818 -0.0009915946
#>  [4,]  0.0004476547  2.319123e-05  0.0005176117  0.0005751141  0.0006324002
#>  [5,] -0.0008178497 -1.862020e-04 -0.0013327411 -0.0014774185 -0.0013210270
#>  [6,]  0.0008126159  7.126361e-05  0.0010250677  0.0011409199  0.0011814296
#>  [7,] -0.0012992514 -2.501992e-04 -0.0019666164 -0.0021466323 -0.0019665253
#>  [8,]  0.0001625836 -9.461275e-06  0.0001196306  0.0001220957  0.0001710629
#>  [9,]  0.0009647481  9.275404e-05  0.0012266460  0.0013532741  0.0013598815
#> [10,]  0.0003713637  5.530514e-06  0.0004186908  0.0004738677  0.0005352739
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0010821789  0.0010031465  0.0008986466  0.0008816931  0.0010085266
#>  [2,]  0.0004797426  0.0004858508  0.0004978722  0.0004930168  0.0004932247
#>  [3,] -0.0007434403 -0.0005239258 -0.0002242166 -0.0002089763 -0.0004829962
#>  [4,]  0.0006367968  0.0006421130  0.0006552684  0.0006440335  0.0006681465
#>  [5,] -0.0010294768 -0.0007835605 -0.0004336855 -0.0004061758 -0.0007280667
#>  [6,]  0.0010825724  0.0010070652  0.0009270574  0.0008902467  0.0010366281
#>  [7,] -0.0015377587 -0.0011751333 -0.0007267163 -0.0007006417 -0.0011683605
#>  [8,]  0.0002193874  0.0002462458  0.0002836971  0.0002884746  0.0002581761
#>  [9,]  0.0012117693  0.0010761260  0.0009107946  0.0008598999  0.0010615383
#> [10,]  0.0005422193  0.0005707508  0.0006268044  0.0006111063  0.0006195735
#>                 V28           V29            V3           V30          V31
#>  [1,]  0.0011434806  0.0010032162  6.343955e-05  0.0007765620 0.0003741516
#>  [2,]  0.0004865776  0.0004205840  1.374977e-05  0.0003414077 0.0002781973
#>  [3,] -0.0007899039 -0.0007205957 -9.540929e-05 -0.0004614803 0.0002732935
#>  [4,]  0.0006775349  0.0005724830  1.646151e-05  0.0004568291 0.0003399682
#>  [5,] -0.0011322102 -0.0010370060 -1.182092e-04 -0.0007391102 0.0001321027
#>  [6,]  0.0011506475  0.0009700933  4.341556e-05  0.0007163002 0.0003173714
#>  [7,] -0.0016732513 -0.0014635419 -1.581849e-04 -0.0009682474 0.0003126600
#>  [8,]  0.0002053680  0.0001748809 -1.078501e-06  0.0001465938 0.0001999002
#>  [9,]  0.0012110087  0.0010222497  6.018312e-05  0.0007029204 0.0001583691
#> [10,]  0.0006157807  0.0005069199 -7.595038e-07  0.0004213674 0.0003700542
#>                V32           V33           V34           V35           V36
#>  [1,] 3.019467e-04  1.374397e-04 -1.146412e-04 -0.0001637579 -2.685222e-04
#>  [2,] 2.067649e-04  1.492039e-04  9.651805e-05  0.0001027583  8.383858e-05
#>  [3,] 1.624062e-04  3.115208e-04  7.190331e-04  0.0009010850  1.027054e-03
#>  [4,] 2.406560e-04  1.593385e-04  8.722456e-05  0.0001195347  8.449460e-05
#>  [5,] 4.201917e-05  2.365607e-04  7.347713e-04  0.0009315110  1.151559e-03
#>  [6,] 2.110707e-04  4.009852e-05 -1.795163e-04 -0.0001527492 -2.410648e-04
#>  [7,] 2.191328e-04  5.402557e-04  1.243672e-03  0.0014647753  1.677189e-03
#>  [8,] 1.464920e-04  1.231804e-04  1.358445e-04  0.0001469357  1.702858e-04
#>  [9,] 9.585647e-05 -7.566690e-05 -3.608031e-04 -0.0003796213 -4.666773e-04
#> [10,] 2.437585e-04  1.730260e-04  1.513499e-04  0.0002344087  1.997672e-04
#>                 V37           V38           V39            V4           V40
#>  [1,] -1.493896e-04  9.538488e-05  2.410147e-04  1.106989e-04  2.773592e-04
#>  [2,]  5.183051e-05  3.855476e-05  7.740247e-05  1.782435e-05  1.405077e-04
#>  [3,]  5.932300e-04 -4.575076e-05 -2.726704e-04 -1.841552e-04 -1.190040e-04
#>  [4,]  4.962605e-05  5.307127e-05  9.428168e-05  2.783510e-05  1.613702e-04
#>  [5,]  6.949532e-04 -4.631777e-05 -3.103933e-04 -2.294861e-04 -1.596312e-04
#>  [6,] -1.530625e-04  6.177699e-05  1.793678e-04  8.920618e-05  2.191926e-04
#>  [7,]  9.798316e-04 -6.375623e-05 -4.516378e-04 -3.102303e-04 -2.492192e-04
#>  [8,]  1.110679e-04  3.292817e-05  4.195061e-05 -1.154091e-05  9.786284e-05
#>  [9,] -2.812749e-04  6.521684e-05  2.413508e-04  1.233971e-04  2.429099e-04
#> [10,]  1.071398e-04  4.522201e-05  4.402268e-05  5.189620e-06  1.200657e-04
#>                 V41           V42           V43           V44           V45
#>  [1,]  3.894941e-04  5.467639e-04  5.913890e-04  6.068370e-04  8.504880e-04
#>  [2,]  1.346654e-04  1.234746e-04  1.413453e-04  1.214969e-04  1.413335e-04
#>  [3,] -3.786831e-04 -8.108617e-04 -8.490580e-04 -9.459798e-04 -1.422380e-03
#>  [4,]  1.673270e-04  1.610345e-04  1.977626e-04  1.827089e-04  2.272382e-04
#>  [5,] -5.085829e-04 -1.024042e-03 -1.057721e-03 -1.147057e-03 -1.752947e-03
#>  [6,]  3.120472e-04  4.241410e-04  5.040666e-04  5.245403e-04  7.227592e-04
#>  [7,] -6.824208e-04 -1.369464e-03 -1.465417e-03 -1.633810e-03 -2.455980e-03
#>  [8,]  5.166387e-05 -1.099004e-05 -8.885063e-06 -3.163579e-05 -8.698179e-05
#>  [9,]  3.403781e-04  5.227518e-04  6.270332e-04  6.645655e-04  9.106515e-04
#> [10,]  1.078203e-04  4.575128e-05  9.058543e-05  7.345518e-05  7.615093e-05
#>                 V46           V47           V48           V49            V5
#>  [1,]  6.733401e-04  4.182262e-04  2.800681e-04  1.747520e-04  1.808054e-04
#>  [2,]  1.124311e-04  6.996786e-05  3.860955e-05  2.518454e-05  5.038766e-05
#>  [3,] -1.124052e-03 -7.016066e-04 -5.006457e-04 -3.086256e-04 -2.248785e-04
#>  [4,]  1.801937e-04  1.070236e-04  5.893681e-05  3.994368e-05  7.072903e-05
#>  [5,] -1.400086e-03 -8.631977e-04 -6.095821e-04 -3.743911e-04 -2.879100e-04
#>  [6,]  5.664308e-04  3.426453e-04  2.195016e-04  1.446576e-04  1.649366e-04
#>  [7,] -1.945343e-03 -1.201100e-03 -8.350256e-04 -5.213546e-04 -4.049896e-04
#>  [8,] -7.148611e-05 -4.012447e-05 -3.299605e-05 -1.960887e-05  6.762647e-06
#>  [9,]  7.024116e-04  4.371407e-04  2.948430e-04  1.951839e-04  2.065088e-04
#> [10,]  6.087079e-05  2.444845e-05 -1.595832e-06  4.775746e-06  4.481047e-05
#>                 V50           V51           V52           V53           V54
#>  [1,]  3.204772e-05  3.888600e-05  3.822686e-05  1.842999e-05  1.998116e-05
#>  [2,]  8.900342e-06  4.753693e-06  5.531566e-06  6.072021e-06  4.194646e-06
#>  [3,] -3.822362e-05 -6.956880e-05 -6.531028e-05 -2.055564e-05 -3.078281e-05
#>  [4,]  1.177221e-05  8.058385e-06  9.628607e-06  7.244546e-06  5.620973e-06
#>  [5,] -5.112785e-05 -8.606159e-05 -8.240828e-05 -2.801958e-05 -3.987486e-05
#>  [6,]  2.442725e-05  3.074911e-05  3.233785e-05  1.384911e-05  1.485900e-05
#>  [7,] -6.604527e-05 -1.148225e-04 -1.103777e-04 -3.469216e-05 -5.109342e-05
#>  [8,]  4.676815e-07 -5.229304e-06 -5.203855e-06  1.440883e-06 -1.132747e-06
#>  [9,]  2.700748e-05  4.140728e-05  4.150115e-05  1.616559e-05  1.929392e-05
#> [10,]  7.027787e-06 -1.263518e-07  3.063023e-06  3.860672e-06  1.512193e-06
#>                 V55           V56           V57           V58           V59
#>  [1,]  1.762272e-05  8.034096e-06 -5.645953e-07  1.175211e-05  1.075150e-05
#>  [2,]  4.525666e-06  4.471691e-06  2.855149e-06  1.787874e-06  2.594347e-06
#>  [3,] -2.218211e-05 -2.681748e-06  1.311919e-05 -1.991551e-05 -1.521128e-05
#>  [4,]  6.724460e-06  4.639170e-06  2.326103e-06  2.717855e-06  3.486433e-06
#>  [5,] -3.035779e-05 -5.163162e-06  1.341066e-05 -2.532929e-05 -1.990733e-05
#>  [6,]  1.513634e-05  5.341305e-06 -2.921677e-06  9.159932e-06  8.393049e-06
#>  [7,] -3.934793e-05 -5.166119e-06  2.215541e-05 -3.263978e-05 -2.569451e-05
#>  [8,] -3.634674e-07  2.878762e-06  3.490865e-06 -1.291679e-06 -1.734464e-07
#>  [9,]  1.775255e-05  5.211854e-06 -5.985359e-06  1.257396e-05  1.101213e-05
#> [10,]  4.654691e-06  3.344915e-06  2.685243e-06  2.433006e-07  1.284575e-06
#>                  V6           V60            V7            V8            V9
#>  [1,]  1.312975e-04  8.975057e-06  1.853306e-04  2.229164e-04  3.480642e-04
#>  [2,]  6.623314e-05  4.001799e-06  8.035839e-05  7.763811e-05  8.837138e-05
#>  [3,] -4.122435e-05 -6.000060e-06 -1.145629e-04 -2.177874e-04 -4.726136e-04
#>  [4,]  8.186119e-05  4.782035e-06  1.047106e-04  1.025767e-04  1.221175e-04
#>  [5,] -7.546936e-05 -9.420227e-06 -1.684417e-04 -2.853658e-04 -5.778767e-04
#>  [6,]  1.185780e-04  7.014799e-06  1.690526e-04  1.908978e-04  2.876620e-04
#>  [7,] -1.116569e-04 -1.203002e-05 -2.506458e-04 -4.211927e-04 -8.505476e-04
#>  [8,]  4.292710e-05  1.713116e-06  3.834240e-05  2.668035e-05  6.369906e-06
#>  [9,]  1.208594e-04  7.611521e-06  1.824249e-04  2.193836e-04  3.529361e-04
#> [10,]  7.335784e-05  3.509995e-06  9.144058e-05  7.569554e-05  6.636302e-05
#> 
#> $vW[[2]]
#>              [,1]        [,2]         [,3]         [,4]         [,5]
#> [1,] -0.008795138 -0.01199891 -0.006342708 -0.006035324 -0.009503726
#> [2,]  0.008795138  0.01199891  0.006342708  0.006035324  0.009503726
#>             [,6]         [,7]        [,8]        [,9]         [,10]
#> [1,] -0.00160448 -0.008423574 -0.01484918 -0.00428885 -0.0006926304
#> [2,]  0.00160448  0.008423574  0.01484918  0.00428885  0.0006926304
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.35902228 0.45667023 0.28122448 0.19911140 0.51854079 0.01883039
#>  [7] 0.35584047 0.68794514 0.22874803 0.27158204
#> 
#> $B[[2]]
#> [1]  0.10030152 -0.05537811
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  8.938596e-04  5.390528e-04 -1.745692e-05  6.681559e-04 -2.499233e-04
#>  [6]  8.230084e-04 -3.265750e-04  3.530148e-04  7.669450e-04  6.329176e-04
#> 
#> $vB[[2]]
#> [1] -0.0194383  0.0194383
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#>  [2,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#>  [3,] 0.0116 0.1006 0.2500 0.3988 0.3809 0.4753 0.6165 0.6464 0.8024 0.9208
#>  [4,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#>  [5,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#>  [6,] 0.0164 0.0780 0.1791 0.2681 0.1788 0.1039 0.1980 0.3234 0.3748 0.2586
#>  [7,] 0.0522 0.2529 0.2716 0.2374 0.1878 0.0983 0.0683 0.1503 0.1723 0.2339
#>  [8,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#>  [9,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [10,] 0.0200 0.2111 0.1609 0.1582 0.2238 0.0645 0.0660 0.2273 0.3100 0.2999
#> [11,] 0.0286 0.3039 0.2988 0.4250 0.6343 0.8198 1.0000 0.9988 0.9508 0.9025
#> [12,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#> [13,] 0.0209 0.3914 0.3504 0.3669 0.3943 0.3311 0.3331 0.3002 0.2324 0.1381
#> [14,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [15,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [16,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [17,] 0.0233 0.2078 0.1239 0.0236 0.1771 0.3115 0.4990 0.6707 0.7655 0.8485
#> [18,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [19,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [20,] 0.0071 0.0898 0.0289 0.1554 0.1437 0.1035 0.1424 0.1227 0.0892 0.2047
#> [21,] 0.1083 0.5966 0.5304 0.2251 0.2402 0.2689 0.6646 0.6632 0.1674 0.0837
#> [22,] 0.0131 0.2079 0.2295 0.1990 0.1184 0.1891 0.2949 0.5343 0.6850 0.7923
#> [23,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [24,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [25,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [26,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [27,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [28,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [29,] 0.0131 0.0901 0.0750 0.0844 0.1226 0.1619 0.2317 0.2934 0.3526 0.3657
#> [30,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [31,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [32,] 0.0130 0.1375 0.2026 0.2389 0.2112 0.1444 0.0742 0.1533 0.3052 0.4116
#> [33,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [34,] 0.0283 0.1767 0.1995 0.2869 0.3275 0.3769 0.4169 0.5036 0.6180 0.8025
#> [35,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [36,] 0.0132 0.0922 0.1445 0.1475 0.2087 0.2558 0.2603 0.1985 0.2394 0.3134
#> [37,] 0.0291 0.1146 0.0942 0.0794 0.0252 0.1191 0.1045 0.2050 0.1556 0.2690
#> [38,] 0.0305 0.3127 0.2192 0.2621 0.2419 0.2179 0.1159 0.1237 0.0886 0.1755
#> [39,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [40,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#>  [2,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#>  [3,] 0.9832 0.0179 0.9634 0.8646 0.8325 0.8276 0.8007 0.6102 0.4853 0.4355
#>  [4,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#>  [5,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#>  [6,] 0.3680 0.0627 0.3508 0.5606 0.5231 0.5469 0.6954 0.6352 0.6757 0.8499
#>  [7,] 0.1962 0.0437 0.1395 0.3164 0.5888 0.7631 0.8473 0.9424 0.9986 0.9699
#>  [8,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#>  [9,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [10,] 0.5078 0.0371 0.4797 0.5783 0.5071 0.4328 0.5550 0.6711 0.6415 0.7104
#> [11,] 0.7234 0.0453 0.5122 0.2074 0.3985 0.5890 0.2872 0.2043 0.5782 0.5389
#> [12,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#> [13,] 0.3450 0.0191 0.4428 0.4890 0.3677 0.4379 0.4864 0.6207 0.7256 0.6624
#> [14,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [15,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [16,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [17,] 0.9805 0.0394 1.0000 1.0000 0.9992 0.9067 0.6803 0.5103 0.4716 0.4980
#> [18,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [19,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [20,] 0.0827 0.0103 0.1524 0.3031 0.1608 0.0667 0.1426 0.0395 0.1653 0.3399
#> [21,] 0.4331 0.1070 0.8718 0.7992 0.3712 0.1703 0.1611 0.2086 0.2847 0.2211
#> [22,] 0.8220 0.0201 0.7290 0.7352 0.7918 0.8057 0.4898 0.1934 0.2924 0.6255
#> [23,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [24,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [25,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [26,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [27,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [28,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [29,] 0.3221 0.0068 0.3093 0.4084 0.4285 0.4663 0.5956 0.6948 0.8386 0.8875
#> [30,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [31,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [32,] 0.5466 0.0120 0.5933 0.6663 0.7333 0.7136 0.7014 0.7758 0.9137 0.9964
#> [33,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [34,] 0.9333 0.0599 0.9399 0.9275 0.9450 0.8328 0.7773 0.7007 0.6154 0.5810
#> [35,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [36,] 0.4077 0.0080 0.4529 0.4893 0.5666 0.6234 0.6741 0.8282 0.8823 0.9196
#> [37,] 0.3784 0.0400 0.4024 0.3470 0.1395 0.1208 0.2827 0.1500 0.2626 0.4468
#> [38,] 0.1758 0.0363 0.1540 0.0512 0.1805 0.4039 0.5697 0.6577 0.7474 0.8543
#> [39,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [40,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#>  [2,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#>  [3,] 0.4307 0.4399 0.0449 0.3833 0.3032 0.3035 0.3197 0.2292 0.2131 0.2347
#>  [4,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#>  [5,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#>  [6,] 0.8025 0.6563 0.0738 0.8591 0.6655 0.5369 0.3118 0.3763 0.2801 0.0875
#>  [7,] 1.0000 0.8630 0.0180 0.6979 0.7717 0.7305 0.5197 0.1786 0.1098 0.1446
#>  [8,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#>  [9,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [10,] 0.8080 0.6791 0.0428 0.3857 0.1307 0.2604 0.5121 0.7547 0.8537 0.8507
#> [11,] 0.3750 0.3411 0.0277 0.5067 0.5580 0.4778 0.3299 0.2198 0.1407 0.2856
#> [12,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#> [13,] 0.7689 0.7981 0.0411 0.8577 0.9273 0.7009 0.4851 0.3409 0.1406 0.1147
#> [14,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [15,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [16,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [17,] 0.6196 0.7171 0.0416 0.6316 0.3554 0.2897 0.4316 0.3791 0.2421 0.0944
#> [18,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [19,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [20,] 0.4855 0.5206 0.0135 0.5508 0.6102 0.5989 0.6764 0.8897 1.0000 0.9517
#> [21,] 0.6134 0.5807 0.0257 0.6925 0.3825 0.4303 0.7791 0.8703 1.0000 0.9212
#> [22,] 0.8546 0.8966 0.0045 0.7821 0.5168 0.4840 0.4038 0.3411 0.2849 0.2353
#> [23,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [24,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [25,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [26,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [27,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [28,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [29,] 0.6404 0.3308 0.0308 0.3425 0.4920 0.4592 0.3034 0.4366 0.5175 0.5122
#> [30,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [31,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [32,] 1.0000 0.8881 0.0436 0.6585 0.2707 0.1746 0.2709 0.4853 0.7184 0.8209
#> [33,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [34,] 0.4454 0.3707 0.0656 0.2891 0.2185 0.1711 0.3578 0.3947 0.2867 0.2401
#> [35,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [36,] 0.8965 0.7549 0.0188 0.6736 0.6463 0.5007 0.3663 0.2298 0.1362 0.2123
#> [37,] 0.7520 0.9036 0.0771 0.7812 0.4766 0.2483 0.5372 0.6279 0.3647 0.4572
#> [38,] 0.9085 0.8668 0.0214 0.8892 0.9065 0.8522 0.7204 0.6200 0.6253 0.6848
#> [39,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [40,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#>  [2,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#>  [3,] 0.3201 0.4455 0.3655 0.1096 0.2715 0.1747 0.1781 0.2199 0.1056 0.0573
#>  [4,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#>  [5,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#>  [6,] 0.3319 0.4237 0.1801 0.0608 0.3743 0.4627 0.1614 0.2494 0.3202 0.2265
#>  [7,] 0.1066 0.1440 0.1929 0.0292 0.0325 0.1490 0.0328 0.0537 0.1309 0.0910
#>  [8,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#>  [9,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [10,] 0.6692 0.6097 0.4943 0.0207 0.2744 0.0510 0.2834 0.2825 0.4256 0.2641
#> [11,] 0.3807 0.4158 0.4054 0.0174 0.3296 0.2707 0.2650 0.0723 0.1238 0.1192
#> [12,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#> [13,] 0.1433 0.1820 0.3605 0.0321 0.5529 0.5988 0.5077 0.5512 0.5027 0.7034
#> [14,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [15,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [16,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [17,] 0.0351 0.0844 0.0436 0.0547 0.1130 0.2045 0.1937 0.0834 0.1502 0.1675
#> [18,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [19,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [20,] 0.8459 0.7073 0.6697 0.0494 0.6326 0.5102 0.4161 0.2816 0.1705 0.1421
#> [21,] 0.9386 0.9303 0.7314 0.0837 0.4791 0.2087 0.2016 0.1669 0.2872 0.4374
#> [22,] 0.2699 0.4442 0.4323 0.0217 0.3314 0.1195 0.1669 0.3702 0.3072 0.0945
#> [23,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [24,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [25,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [26,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [27,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [28,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [29,] 0.4746 0.4902 0.4603 0.0311 0.4460 0.4196 0.2873 0.2296 0.0949 0.0095
#> [30,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [31,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [32,] 0.7536 0.6496 0.4708 0.0624 0.3482 0.3508 0.3181 0.3524 0.3659 0.2846
#> [33,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [34,] 0.3619 0.3314 0.3763 0.0229 0.4767 0.4059 0.3661 0.2320 0.1450 0.1017
#> [35,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [36,] 0.2395 0.2673 0.2865 0.0141 0.2060 0.1659 0.2633 0.2552 0.1696 0.1467
#> [37,] 0.6359 0.6474 0.5520 0.0809 0.3253 0.2292 0.0653 0.0000 0.0000 0.0000
#> [38,] 0.7337 0.6281 0.5725 0.0227 0.6119 0.5597 0.4965 0.5027 0.5772 0.5907
#> [39,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [40,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#>  [2,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#>  [3,] 0.0307 0.0237 0.0470 0.0102 0.1913 0.0057 0.0031 0.0163 0.0099 0.0084
#>  [4,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#>  [5,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#>  [6,] 0.1146 0.0476 0.0943 0.0824 0.0233 0.0171 0.0244 0.0258 0.0143 0.0226
#>  [7,] 0.0757 0.1059 0.1005 0.0535 0.0351 0.0235 0.0155 0.0160 0.0029 0.0051
#>  [8,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#>  [9,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [10,] 0.1386 0.1051 0.1343 0.0383 0.0954 0.0324 0.0232 0.0027 0.0065 0.0159
#> [11,] 0.1089 0.0623 0.0494 0.0264 0.0384 0.0081 0.0104 0.0045 0.0014 0.0038
#> [12,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#> [13,] 0.5904 0.4069 0.2761 0.1584 0.0698 0.0510 0.0054 0.0078 0.0201 0.0104
#> [14,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [15,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [16,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [17,] 0.1058 0.1111 0.0849 0.0596 0.0993 0.0201 0.0071 0.0104 0.0062 0.0026
#> [18,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [19,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [20,] 0.0971 0.0879 0.0863 0.0355 0.0253 0.0233 0.0252 0.0043 0.0048 0.0076
#> [21,] 0.3097 0.1578 0.0553 0.0334 0.0748 0.0209 0.0172 0.0180 0.0110 0.0234
#> [22,] 0.1545 0.1394 0.0772 0.0615 0.0230 0.0230 0.0111 0.0168 0.0086 0.0045
#> [23,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [24,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [25,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [26,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [27,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [28,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [29,] 0.0527 0.0383 0.0107 0.0108 0.0085 0.0077 0.0109 0.0062 0.0028 0.0040
#> [30,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [31,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [32,] 0.1714 0.0694 0.0303 0.0292 0.0428 0.0116 0.0024 0.0084 0.0100 0.0018
#> [33,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [34,] 0.1111 0.0655 0.0271 0.0244 0.0839 0.0179 0.0109 0.0147 0.0170 0.0158
#> [35,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [36,] 0.1286 0.0926 0.0716 0.0325 0.0436 0.0258 0.0136 0.0044 0.0028 0.0021
#> [37,] 0.0000 0.0000 0.0000 0.0000 0.0521 0.0000 0.0000 0.0056 0.0237 0.0204
#> [38,] 0.4803 0.3877 0.2779 0.1427 0.0456 0.0424 0.0271 0.0200 0.0070 0.0070
#> [39,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [40,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#>  [2,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#>  [3,] 0.0270 0.0277 0.0097 0.0054 0.0148 0.0924 0.0092 0.0761 0.1092 0.0757
#>  [4,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#>  [5,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#>  [6,] 0.0187 0.0185 0.0110 0.0094 0.0078 0.1048 0.0112 0.1338 0.0644 0.1522
#>  [7,] 0.0062 0.0089 0.0140 0.0138 0.0077 0.1171 0.0031 0.1257 0.1178 0.1258
#>  [8,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#>  [9,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [10,] 0.0072 0.0167 0.0180 0.0084 0.0090 0.0986 0.0032 0.1539 0.1601 0.3109
#> [11,] 0.0013 0.0089 0.0057 0.0027 0.0051 0.0990 0.0062 0.1201 0.1833 0.2105
#> [12,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> [13,] 0.0039 0.0031 0.0062 0.0087 0.0070 0.1579 0.0042 0.1438 0.1402 0.3048
#> [14,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [15,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [16,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [17,] 0.0025 0.0061 0.0038 0.0101 0.0078 0.1515 0.0006 0.1674 0.1513 0.1723
#> [18,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [19,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [20,] 0.0124 0.0105 0.0054 0.0032 0.0073 0.0806 0.0063 0.0701 0.0738 0.0117
#> [21,] 0.0276 0.0032 0.0084 0.0122 0.0082 0.1125 0.0143 0.3322 0.4590 0.5526
#> [22,] 0.0062 0.0065 0.0030 0.0066 0.0029 0.0481 0.0053 0.0742 0.0333 0.1369
#> [23,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [24,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [25,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [26,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [27,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [28,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [29,] 0.0075 0.0039 0.0053 0.0013 0.0052 0.0767 0.0023 0.0771 0.0640 0.0726
#> [30,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [31,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [32,] 0.0035 0.0058 0.0011 0.0009 0.0033 0.0349 0.0026 0.0384 0.0446 0.1318
#> [33,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [34,] 0.0046 0.0073 0.0054 0.0033 0.0045 0.1673 0.0079 0.1154 0.1098 0.1370
#> [35,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [36,] 0.0022 0.0048 0.0138 0.0140 0.0028 0.0668 0.0064 0.0609 0.0131 0.0899
#> [37,] 0.0050 0.0137 0.0164 0.0081 0.0139 0.1051 0.0111 0.0145 0.0674 0.1294
#> [38,] 0.0086 0.0089 0.0074 0.0042 0.0055 0.0665 0.0021 0.0939 0.0972 0.2535
#> [39,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [40,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7127074 0.7893993 0.6373489 0.6441412 0.7255546 0.5439554 0.6562503
#>  [2,] 0.6873808 0.7707803 0.6384128 0.6750888 0.6951347 0.5526678 0.6553018
#>  [3,] 0.7117069 0.7646116 0.6284146 0.6429712 0.7110770 0.5370866 0.6523137
#>  [4,] 0.7412351 0.8042037 0.6785026 0.6361015 0.7659752 0.5831748 0.6655216
#>  [5,] 0.7367052 0.7733676 0.6105076 0.5497229 0.7462347 0.5127731 0.5725831
#>  [6,] 0.6752627 0.7512229 0.6427693 0.6223125 0.7292868 0.5083147 0.6422096
#>  [7,] 0.7099947 0.8066789 0.6717650 0.6614321 0.7680754 0.6119040 0.6884133
#>  [8,] 0.6770728 0.7093435 0.5537183 0.5500773 0.6694888 0.4239614 0.5428483
#>  [9,] 0.7318532 0.7765558 0.6432138 0.6373511 0.7268439 0.5745080 0.6092863
#> [10,] 0.7128532 0.7267225 0.6087404 0.5192504 0.7257658 0.4310286 0.5592649
#> [11,] 0.6085574 0.7144589 0.5520604 0.5990186 0.6500413 0.4635508 0.6635936
#> [12,] 0.7419159 0.8251368 0.7007917 0.7073298 0.7610414 0.6376919 0.6937510
#> [13,] 0.5736916 0.6722512 0.5334334 0.5312800 0.6261379 0.3566557 0.5856420
#> [14,] 0.7601379 0.8148916 0.6875451 0.6341005 0.7994792 0.5779540 0.6519706
#> [15,] 0.6327871 0.6381939 0.5116741 0.5043983 0.6463455 0.3172163 0.5616474
#> [16,] 0.6964827 0.7984153 0.6604370 0.6503061 0.7493332 0.5616283 0.6563015
#> [17,] 0.7059602 0.7928041 0.6587067 0.6915523 0.7203048 0.5821492 0.6729063
#> [18,] 0.6245167 0.7210562 0.5826850 0.5957530 0.6743221 0.4307405 0.6007621
#> [19,] 0.7471875 0.7935320 0.6639907 0.6443728 0.7643488 0.5453800 0.6429143
#> [20,] 0.6486350 0.6437242 0.5448807 0.5012950 0.6770843 0.3638253 0.5270969
#> [21,] 0.6081719 0.6267425 0.4840249 0.4347682 0.6272902 0.2873835 0.4567796
#> [22,] 0.7124623 0.7727362 0.6530232 0.6621022 0.7293930 0.5277222 0.6866417
#> [23,] 0.6009933 0.6930508 0.5355463 0.5413392 0.6356623 0.4032950 0.5468713
#> [24,] 0.6990504 0.7896827 0.6550819 0.6788986 0.7210677 0.5475001 0.6890607
#> [25,] 0.7625211 0.7856470 0.6503516 0.5588031 0.7932123 0.4849650 0.5883023
#> [26,] 0.7216701 0.7711716 0.5966124 0.5707955 0.7289311 0.4977827 0.5748631
#> [27,] 0.6724499 0.7647049 0.6065039 0.5709142 0.7251719 0.4538948 0.6338808
#> [28,] 0.7408923 0.8222932 0.7094692 0.6886145 0.7806194 0.6444142 0.6952861
#> [29,] 0.7170929 0.7600898 0.6478094 0.6072157 0.7570808 0.5186048 0.6405015
#> [30,] 0.7425805 0.8237032 0.7014259 0.6857790 0.7741537 0.6005341 0.7025905
#> [31,] 0.7359066 0.8033774 0.6642993 0.6508872 0.7682147 0.5527621 0.6668284
#> [32,] 0.7676204 0.7901590 0.6757315 0.5931507 0.8046802 0.5007085 0.6283396
#> [33,] 0.7233663 0.8049729 0.6895957 0.6503782 0.7708571 0.5851611 0.6788763
#> [34,] 0.7161966 0.7598411 0.6213020 0.6181583 0.7229048 0.5236057 0.6243047
#> [35,] 0.7528691 0.8170374 0.7085595 0.6626123 0.7884692 0.6169360 0.6896191
#> [36,] 0.7156933 0.7956529 0.6863106 0.6538666 0.7690832 0.5740571 0.6890142
#> [37,] 0.6749088 0.6980081 0.6326076 0.5844773 0.7345423 0.4631458 0.6245551
#> [38,] 0.6247601 0.6767272 0.5430077 0.4974099 0.6841591 0.3360882 0.5627573
#> [39,] 0.7438127 0.7944850 0.6537127 0.6350330 0.7632087 0.5582517 0.5946755
#> [40,] 0.6852159 0.7498461 0.6467897 0.6014163 0.7410980 0.5317417 0.6465296
#>            [,8]      [,9]     [,10]
#>  [1,] 0.8478565 0.5610811 0.5456274
#>  [2,] 0.8225206 0.5396535 0.5177412
#>  [3,] 0.8186688 0.5596918 0.5313737
#>  [4,] 0.8738276 0.5567607 0.5915678
#>  [5,] 0.8518076 0.4424587 0.4541315
#>  [6,] 0.8428916 0.4362480 0.5464343
#>  [7,] 0.8761946 0.4911349 0.6424575
#>  [8,] 0.7841479 0.3306935 0.3612064
#>  [9,] 0.8340317 0.6035407 0.5623562
#> [10,] 0.8195508 0.3888128 0.4604241
#> [11,] 0.7568612 0.4483455 0.4777322
#> [12,] 0.8764166 0.6349105 0.6345823
#> [13,] 0.7684254 0.2772401 0.3881924
#> [14,] 0.8896161 0.5138832 0.5997942
#> [15,] 0.7423624 0.2518058 0.3423367
#> [16,] 0.8700775 0.4053674 0.5684523
#> [17,] 0.8490615 0.5738677 0.5486268
#> [18,] 0.8083892 0.3254627 0.4423597
#> [19,] 0.8709388 0.5629426 0.5754455
#> [20,] 0.7486071 0.2733469 0.3741816
#> [21,] 0.7035410 0.2535359 0.2450824
#> [22,] 0.8467491 0.4834755 0.5422487
#> [23,] 0.7967466 0.2314451 0.3742809
#> [24,] 0.8461914 0.5194458 0.5442533
#> [25,] 0.8774877 0.4190660 0.5254788
#> [26,] 0.8550429 0.4389696 0.4673232
#> [27,] 0.8475426 0.3952112 0.4707072
#> [28,] 0.8888031 0.5763563 0.6805270
#> [29,] 0.8442877 0.4864566 0.5786023
#> [30,] 0.8848823 0.5705087 0.6180118
#> [31,] 0.8734058 0.5666532 0.5830018
#> [32,] 0.8828195 0.4603950 0.5534195
#> [33,] 0.8768887 0.5391315 0.6090470
#> [34,] 0.8334022 0.5543765 0.5158853
#> [35,] 0.8876709 0.5837059 0.6237676
#> [36,] 0.8704532 0.5118639 0.6215917
#> [37,] 0.7939812 0.4249394 0.5007672
#> [38,] 0.7985886 0.2124041 0.4007670
#> [39,] 0.8781258 0.3993104 0.5245600
#> [40,] 0.8291690 0.5226091 0.6136228
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.5614031 0.4385969
#>  [2,] 0.5609503 0.4390497
#>  [3,] 0.5607159 0.4392841
#>  [4,] 0.5630442 0.4369558
#>  [5,] 0.5613168 0.4386832
#>  [6,] 0.5648347 0.4351653
#>  [7,] 0.5648997 0.4351003
#>  [8,] 0.5609650 0.4390350
#>  [9,] 0.5579175 0.4420825
#> [10,] 0.5628449 0.4371551
#> [11,] 0.5626650 0.4373350
#> [12,] 0.5615942 0.4384058
#> [13,] 0.5645652 0.4354348
#> [14,] 0.5644481 0.4355519
#> [15,] 0.5642675 0.4357325
#> [16,] 0.5656923 0.4343077
#> [17,] 0.5614223 0.4385777
#> [18,] 0.5645368 0.4354632
#> [19,] 0.5621178 0.4378822
#> [20,] 0.5630997 0.4369003
#> [21,] 0.5600170 0.4399830
#> [22,] 0.5649268 0.4350732
#> [23,] 0.5630880 0.4369120
#> [24,] 0.5638474 0.4361526
#> [25,] 0.5646912 0.4353088
#> [26,] 0.5608885 0.4391115
#> [27,] 0.5657818 0.4342182
#> [28,] 0.5635272 0.4364728
#> [29,] 0.5639479 0.4360521
#> [30,] 0.5644813 0.4355187
#> [31,] 0.5631781 0.4368219
#> [32,] 0.5659426 0.4340574
#> [33,] 0.5644848 0.4355152
#> [34,] 0.5602726 0.4397274
#> [35,] 0.5640307 0.4359693
#> [36,] 0.5655326 0.4344674
#> [37,] 0.5653293 0.4346707
#> [38,] 0.5667996 0.4332004
#> [39,] 0.5631398 0.4368602
#> [40,] 0.5630661 0.4369339
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]      [,2]        [,3]         [,4]      [,5]         [,6]
#>  [1,] 0.9085696 1.3213081  0.56387597  0.593383741 0.9721841  0.176276810
#>  [2,] 0.7879024 1.2127222  0.56848166  0.731292215 0.8242355  0.211455479
#>  [3,] 0.9036883 1.1781312  0.52542124  0.588283279 0.9006207  0.148619385
#>  [4,] 1.0523980 1.4127776  0.74689883  0.558483160 1.1857229  0.335819996
#>  [5,] 1.0289134 1.2274250  0.44944648  0.199551301 1.0786306  0.051103381
#>  [6,] 0.7320852 1.1051450  0.58740411  0.499375244 0.9910070  0.033261854
#>  [7,] 0.8953585 1.4285731  0.71617870  0.669682796 1.1974755  0.455322853
#>  [8,] 0.7403521 0.8921978  0.21570571  0.200983207 0.7058737 -0.306532445
#>  [9,] 1.0040454 1.2457066  0.58934048  0.563885319 0.9786683  0.300267696
#> [10,] 0.9092819 0.9780572  0.44202073  0.077039864 0.9732447 -0.277655608
#> [11,] 0.4412524 0.9171397  0.20899905  0.401377542 0.6192210 -0.146055950
#> [12,] 1.0559506 1.5515455  0.85107053  0.882450806 1.1583975  0.565360506
#> [13,] 0.2969289 0.7183848  0.13393334  0.125283585 0.5156838 -0.589909122
#> [14,] 1.1534359 1.4821134  0.78866717  0.549848713 1.3830428  0.314380169
#> [15,] 0.5441922 0.5675337  0.04670476  0.017593585 0.6030138 -0.766594282
#> [16,] 0.8306043 1.3764192  0.66524240  0.620385145 1.0950589  0.247772885
#> [17,] 0.8758439 1.3419112  0.65753629  0.807386613 0.9459740  0.331602194
#> [18,] 0.5087641 0.9497065  0.33380528  0.387800007 0.7277990 -0.278830610
#> [19,] 1.0836683 1.3463481  0.68112916  0.594394475 1.1766715  0.182021089
#> [20,] 0.6130447 0.5915653  0.18000730  0.005179845 0.7404043 -0.558799360
#> [21,] 0.4396342 0.5182673 -0.06392207 -0.262422836 0.5206093 -0.908125894
#> [22,] 0.9073731 1.2238262  0.63235461  0.672676237 0.9915450  0.111002631
#> [23,] 0.4096057 0.8144213  0.14242540  0.165735331 0.5565863 -0.391754645
#> [24,] 0.8427799 1.3230136  0.64145334  0.748714764 0.9497640  0.190575202
#> [25,] 1.1665512 1.2988834  0.62058496  0.236306058 1.3443984 -0.060158256
#> [26,] 0.9527612 1.2149383  0.39136975  0.285097320 0.9892060 -0.008869156
#> [27,] 0.7192866 1.1786496  0.43264021  0.285581883 0.9702630 -0.184945991
#> [28,] 1.0506113 1.5319620  0.89280756  0.793650045 1.2692794  0.594575287
#> [29,] 0.9300867 1.1531716  0.60942383  0.435623480 1.1367409  0.074453526
#> [30,] 1.0594244 1.5416411  0.85409713  0.780459081 1.2319157  0.407691001
#> [31,] 1.0248005 1.4075382  0.68251301  0.622941204 1.1982575  0.211836891
#> [32,] 1.1949231 1.3258839  0.73422374  0.377005810 1.4158069  0.002833965
#> [33,] 0.9612215 1.4176700  0.79822977  0.620701903 1.2131574  0.343996978
#> [34,] 0.9256727 1.1518086  0.49507830  0.481738517 0.9589164  0.094493190
#> [35,] 1.1139733 1.4964031  0.88839814  0.674957321 1.3157228  0.476563196
#> [36,] 0.9231979 1.3593433  0.78292686  0.636079156 1.2031418  0.298423674
#> [37,] 0.7304717 0.8378304  0.54341992  0.341180567 1.0177922 -0.147684733
#> [38,] 0.5098023 0.7387717  0.17245709 -0.010360541 0.7729518 -0.680775739
#> [39,] 1.0658802 1.3521749  0.63539936  0.553869820 1.1703521  0.234069746
#> [40,] 0.7778467 1.0977915  0.60495765  0.411369879 1.0516832  0.127137736
#>             [,7]      [,8]        [,9]        [,10]
#>  [1,]  0.6466286 1.7178869  0.24555062  0.183018598
#>  [2,]  0.6424266 1.5335188  0.15894768  0.070994763
#>  [3,]  0.6292251 1.5073547  0.23991134  0.125659931
#>  [4,]  0.6879988 1.9352341  0.22802589  0.370450350
#>  [5,]  0.2923981 1.7488489 -0.23118948 -0.183991360
#>  [6,]  0.5849674 1.6799026 -0.25640364  0.186273946
#>  [7,]  0.7927117 1.9568770 -0.03546407  0.586046349
#>  [8,]  0.1718145 1.2900040 -0.70505010 -0.570131680
#>  [9,]  0.4443133 1.6144745  0.42023997  0.250729933
#> [10,]  0.2381793 1.5133071 -0.45230529 -0.158635391
#> [11,]  0.6793500 1.1355474 -0.20735787 -0.089130359
#> [12,]  0.8177146 1.9589252  0.55334133  0.551925933
#> [13,]  0.3459781 1.1994414 -0.95819311 -0.454916686
#> [14,]  0.6277126 2.0868257  0.05554693  0.404607611
#> [15,]  0.2478507 1.0582836 -1.08900439 -0.652898265
#> [16,]  0.6468556 1.9016443 -0.38314997  0.275539533
#> [17,]  0.7213596 1.7272590  0.29764919  0.195124077
#> [18,]  0.4086417 1.4395771 -0.72877890 -0.231590830
#> [19,]  0.5880356 1.9092846  0.25311298  0.304103940
#> [20,]  0.1084937 1.0911974 -0.97770726 -0.514319070
#> [21,] -0.1733140 0.8642173 -1.07984223 -1.125014103
#> [22,]  0.7844655 1.7093277 -0.06612213  0.169398729
#> [23,]  0.1880373 1.3660830 -1.20016904 -0.513894978
#> [24,]  0.7957316 1.7050363  0.07782242  0.177477729
#> [25,]  0.3569516 1.9688514 -0.32660873  0.102003431
#> [26,]  0.3017206 1.7747139 -0.24534497 -0.130893891
#> [27,]  0.5489018 1.7154563 -0.42545901 -0.117305592
#> [28,]  0.8249503 2.0785732  0.30783311  0.756194512
#> [29,]  0.5775414 1.6904830 -0.05418667  0.317038506
#> [30,]  0.8596641 2.0394999  0.28392684  0.481118009
#> [31,]  0.6938751 1.9314134  0.26820905  0.335108514
#> [32,]  0.5251002 2.0194050 -0.15875255  0.214496643
#> [33,]  0.7486125 1.9632909  0.15684666  0.443308152
#> [34,]  0.5078600 1.6099341  0.21836944  0.063562431
#> [35,]  0.7983391 2.0671684  0.33800529  0.505570975
#> [36,]  0.7955148 1.9049716  0.04746433  0.496309715
#> [37,]  0.5089280 1.3490921 -0.30252897  0.003068736
#> [38,]  0.2523599 1.3774964 -1.31049458 -0.402270219
#> [39,]  0.3833278 1.9748002 -0.40833935  0.098319051
#> [40,]  0.6038191 1.5797490  0.09049815  0.462566159
#> 
#> $pre[[3]]
#>            [,1]        [,2]
#>  [1,] 0.1584251 -0.08843329
#>  [2,] 0.1599642 -0.08505553
#>  [3,] 0.1561542 -0.08791366
#>  [4,] 0.1640591 -0.08946693
#>  [5,] 0.1553980 -0.09110980
#>  [6,] 0.1715036 -0.08930355
#>  [7,] 0.1780873 -0.08298430
#>  [8,] 0.1558929 -0.08918634
#>  [9,] 0.1488651 -0.08384941
#> [10,] 0.1577704 -0.09494563
#> [11,] 0.1655408 -0.08644427
#> [12,] 0.1633290 -0.08430537
#> [13,] 0.1679251 -0.09178575
#> [14,] 0.1693284 -0.08990592
#> [15,] 0.1617734 -0.09672654
#> [16,] 0.1797216 -0.08457556
#> [17,] 0.1612406 -0.08569597
#> [18,] 0.1711634 -0.08843212
#> [19,] 0.1591022 -0.09065935
#> [20,] 0.1605124 -0.09323934
#> [21,] 0.1468637 -0.09436726
#> [22,] 0.1699848 -0.09119711
#> [23,] 0.1675394 -0.08616464
#> [24,] 0.1682697 -0.08852162
#> [25,] 0.1651453 -0.09507811
#> [26,] 0.1555291 -0.08923973
#> [27,] 0.1697337 -0.09492768
#> [28,] 0.1722738 -0.08321560
#> [29,] 0.1666449 -0.09055536
#> [30,] 0.1704050 -0.08896442
#> [31,] 0.1631470 -0.09092351
#> [32,] 0.1682247 -0.09709123
#> [33,] 0.1704208 -0.08896275
#> [34,] 0.1524389 -0.08982961
#> [35,] 0.1679101 -0.08962687
#> [36,] 0.1748399 -0.08880733
#> [37,] 0.1682693 -0.09455019
#> [38,] 0.1726494 -0.09615603
#> [39,] 0.1692916 -0.08462301
#> [40,] 0.1657191 -0.08789597
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.4385969 -0.4385969
#>  [2,]  0.4390497 -0.4390497
#>  [3,]  0.4392841 -0.4392841
#>  [4,]  0.4369558 -0.4369558
#>  [5,]  0.4386832 -0.4386832
#>  [6,]  0.4351653 -0.4351653
#>  [7,]  0.4351003 -0.4351003
#>  [8,] -0.5609650  0.5609650
#>  [9,] -0.5579175  0.5579175
#> [10,] -0.5628449  0.5628449
#> [11,] -0.5626650  0.5626650
#> [12,] -0.5615942  0.5615942
#> [13,]  0.4354348 -0.4354348
#> [14,]  0.4355519 -0.4355519
#> [15,] -0.5642675  0.5642675
#> [16,] -0.5656923  0.5656923
#> [17,]  0.4385777 -0.4385777
#> [18,]  0.4354632 -0.4354632
#> [19,]  0.4378822 -0.4378822
#> [20,] -0.5630997  0.5630997
#> [21,]  0.4399830 -0.4399830
#> [22,]  0.4350732 -0.4350732
#> [23,]  0.4369120 -0.4369120
#> [24,] -0.5638474  0.5638474
#> [25,]  0.4353088 -0.4353088
#> [26,] -0.5608885  0.5608885
#> [27,]  0.4342182 -0.4342182
#> [28,] -0.5635272  0.5635272
#> [29,] -0.5639479  0.5639479
#> [30,]  0.4355187 -0.4355187
#> [31,]  0.4368219 -0.4368219
#> [32,]  0.4340574 -0.4340574
#> [33,]  0.4355152 -0.4355152
#> [34,]  0.4397274 -0.4397274
#> [35,]  0.4359693 -0.4359693
#> [36,] -0.5655326  0.5655326
#> [37,] -0.5653293  0.5653293
#> [38,]  0.4332004 -0.4332004
#> [39,] -0.5631398  0.5631398
#> [40,] -0.5630661  0.5630661
#> 
#> $L
#> [1] 0.6895026 0.6707245 0.7143046 0.6934799 0.6969842 0.6753985
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174