Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11         V12         V13          V14
#>  [1,] -0.3410972 -0.12870623 -0.008277792 -0.09418193  0.04217084 -0.051079626
#>  [2,] -0.3568801 -0.19129821 -0.129649519 -0.10722171 -0.18543559 -0.111725149
#>  [3,] -0.3483321 -0.10306876 -0.217403337 -0.05616545 -0.07045549 -0.042797122
#>  [4,] -0.3887752 -0.06600828 -0.090553230  0.01773688 -0.12070170 -0.097113190
#>  [5,] -0.2059574 -0.21289666 -0.092016243 -0.07811718 -0.16119871 -0.164958310
#>  [6,] -0.2536808 -0.25326133 -0.095825357 -0.19952895 -0.11981863 -0.027067072
#>  [7,] -0.3051030 -0.07028269 -0.172559254 -0.10287522  0.03294273  0.009194594
#>  [8,] -0.3469561 -0.10025783 -0.088788796 -0.06594612 -0.05853895 -0.146966995
#>  [9,] -0.2349117 -0.09885870 -0.153023148 -0.03547619 -0.08435456  0.026874300
#> [10,] -0.3025807 -0.10219782 -0.049746826 -0.02954150 -0.08072890 -0.118652500
#>               V15         V16          V17         V18          V19         V2
#>  [1,] -0.01072493  0.03302191 -0.033570946  0.04496856  0.159771281 -0.2108462
#>  [2,] -0.08936407 -0.02758947 -0.010994707 -0.07474210 -0.033864917 -0.4166567
#>  [3,] -0.08491338  0.02182251  0.008484424  0.08537560 -0.002626972 -0.2338661
#>  [4,] -0.12680967 -0.06057951 -0.101912226  0.10295493  0.109532787 -0.2572613
#>  [5,] -0.10451643 -0.16505672 -0.115291513  0.02388452 -0.043029303 -0.3076723
#>  [6,] -0.08017920 -0.01394678  0.020605796 -0.10662686 -0.008521079 -0.1957163
#>  [7,]  0.01296916 -0.06957045  0.009871989  0.02193415  0.083013156 -0.2090646
#>  [8,]  0.02939378 -0.02238546 -0.117628257 -0.05494215 -0.008825674 -0.2669272
#>  [9,] -0.09051641  0.01977144  0.080286028  0.13036635  0.032123649 -0.2624055
#> [10,]  0.07629759  0.08790556  0.066744571 -0.01451274  0.185396473 -0.3192137
#>               V20          V21         V22         V23          V24         V25
#>  [1,]  0.18095814  0.192583012  0.14898789  0.25608303  0.140600016  0.29241676
#>  [2,]  0.04228513 -0.003581935  0.04831216 -0.03398151  0.123694460  0.05273947
#>  [3,]  0.05435877  0.166635945  0.20708074  0.14210736  0.207248774  0.24382330
#>  [4,]  0.04418227  0.137843924  0.16585581  0.09425209  0.165726649  0.04888681
#>  [5,] -0.01941196 -0.048418933 -0.01441839  0.15410998  0.107869722 -0.05297974
#>  [6,]  0.02242719 -0.090348537 -0.11700381 -0.03909338 -0.022018847 -0.04813190
#>  [7,]  0.20040072  0.156176667  0.04493826  0.15077160  0.089450118  0.23485170
#>  [8,] -0.09474281 -0.044628767 -0.05867041  0.07993346 -0.007813733  0.04854350
#>  [9,]  0.07792588  0.198230626  0.29360254  0.15398590  0.306055625  0.16535152
#> [10,]  0.19514253  0.062731416  0.06188338  0.11873358  0.187658482  0.14683377
#>                V26        V27         V28         V29         V3          V30
#>  [1,]  0.236314167 0.30377435  0.19132028  0.09641807 -0.2629062  0.064410038
#>  [2,] -0.037408147 0.19594383  0.08439007  0.07639686 -0.3452770  0.093681889
#>  [3,]  0.260080554 0.17659302  0.19468898  0.10786284 -0.2820438 -0.031276919
#>  [4,]  0.201077387 0.18029269  0.11228031  0.13202136 -0.2095358  0.177293177
#>  [5,]  0.085908392 0.08253182 -0.05496190 -0.07506150 -0.1519710  0.064009717
#>  [6,] -0.002767239 0.02567387 -0.04940397  0.05803054 -0.2815315  0.073994816
#>  [7,]  0.238295919 0.13629685  0.07784406  0.16864089 -0.2311280  0.034388385
#>  [8,] -0.004791769 0.11903027  0.06926604  0.01842459 -0.3100446  0.012971814
#>  [9,]  0.134734796 0.23067721  0.27281357  0.15887502 -0.2360719 -0.009673904
#> [10,]  0.241889673 0.08812013  0.21534209  0.20787805 -0.2923051  0.131217769
#>                V31          V32          V33          V34          V35
#>  [1,]  0.169151683  0.126007552  0.009378181  0.006842614  0.026399310
#>  [2,]  0.064112313  0.045939092  0.012104821  0.098026612 -0.005835002
#>  [3,] -0.055386788 -0.111649500 -0.164761744 -0.182923103 -0.099993967
#>  [4,] -0.018208368 -0.036737125  0.048733537  0.011974791  0.002846610
#>  [5,] -0.053860886 -0.068600029 -0.037843876 -0.061038777 -0.110576640
#>  [6,]  0.007221083 -0.002129832  0.026618428  0.023801915 -0.055577770
#>  [7,]  0.001185475  0.031429086  0.078474554  0.063041496  0.075067030
#>  [8,] -0.040563596  0.002750095 -0.183291174  0.002445821 -0.020088126
#>  [9,] -0.031287441 -0.005772686  0.036460321  0.026039973  0.095901286
#> [10,]  0.123840167 -0.034661917  0.081835820  0.070593820  0.140860387
#>               V36          V37         V38         V39         V4          V40
#>  [1,]  0.05683292 -0.026213337 -0.02945103  0.06825125 -0.3426170  0.039919790
#>  [2,] -0.07455051  0.068563279  0.01547125 -0.06335036 -0.2367491 -0.144054648
#>  [3,] -0.15905228  0.002526606 -0.05371470 -0.06427106 -0.3648621 -0.023808964
#>  [4,]  0.04404492 -0.034812076  0.05595202 -0.12431418 -0.2758582  0.004179531
#>  [5,] -0.05781938 -0.053669685 -0.02071650  0.02964191 -0.2746041 -0.167597002
#>  [6,]  0.10779294 -0.077686631 -0.14419915 -0.11779572 -0.2275945 -0.054092429
#>  [7,]  0.04986132  0.176811357  0.03035528  0.08081653 -0.1316928  0.043907533
#>  [8,]  0.01453985 -0.006875519 -0.15433565 -0.15647363 -0.1913605 -0.188962022
#>  [9,]  0.03479622 -0.062211773 -0.02359473 -0.08187688 -0.2757216  0.011202747
#> [10,]  0.06374589  0.090868968  0.04879915  0.03904466 -0.2541980  0.079571559
#>                V41         V42         V43         V44         V45         V46
#>  [1,]  0.064704697 -0.02506922  0.03653361 -0.13303498 -0.13329352 -0.18458613
#>  [2,] -0.164500242 -0.20291378 -0.13917258 -0.21840576 -0.09014601 -0.16207733
#>  [3,] -0.056720586 -0.13210304 -0.20313101 -0.18056716 -0.25650760 -0.16904980
#>  [4,] -0.134042111 -0.01709118  0.04759133 -0.01855461 -0.08748402 -0.11225598
#>  [5,]  0.019597693 -0.06359147 -0.11532570 -0.13658387 -0.06562872 -0.17750000
#>  [6,] -0.117124309 -0.01303863 -0.02737685 -0.19195558 -0.11306032 -0.24903791
#>  [7,] -0.047275025  0.05469822 -0.09562836 -0.08933195 -0.15902851 -0.14072117
#>  [8,] -0.078533112 -0.02918011 -0.16882537 -0.04077495 -0.16953928 -0.12980451
#>  [9,]  0.002974576  0.01150247 -0.05743911 -0.04114984 -0.01646859 -0.10812285
#> [10,]  0.048788604 -0.06486366 -0.10520952 -0.12830331 -0.14343353 -0.06929808
#>               V47         V48        V49         V5        V50        V51
#>  [1,] -0.17334232 -0.24151275 -0.3154300 -0.1220771 -0.3080436 -0.2190210
#>  [2,] -0.19437870 -0.20855588 -0.3345584 -0.2651643 -0.3095963 -0.3165439
#>  [3,] -0.20636465 -0.24212415 -0.1756222 -0.2282078 -0.2889901 -0.3053788
#>  [4,] -0.25530759 -0.19160807 -0.2763538 -0.1879088 -0.2789323 -0.4146241
#>  [5,] -0.14296604 -0.16023349 -0.2288036 -0.1317555 -0.3764986 -0.3460028
#>  [6,] -0.19462726 -0.22815662 -0.2698977 -0.2646859 -0.2353679 -0.3882676
#>  [7,] -0.04004045 -0.08872156 -0.2360597 -0.2113791 -0.2874806 -0.2231842
#>  [8,] -0.18245765 -0.28715425 -0.1929311 -0.1917602 -0.2512345 -0.3183127
#>  [9,] -0.22757868 -0.15810598 -0.2700956 -0.3183176 -0.3687803 -0.2672442
#> [10,] -0.16608678 -0.22170567 -0.3278500 -0.2942208 -0.3111993 -0.4163753
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.2862226 -0.3489052 -0.2879527 -0.2740841 -0.2532045 -0.3512926
#>  [2,] -0.3842979 -0.4299761 -0.2802196 -0.4009971 -0.4048653 -0.2744975
#>  [3,] -0.3363506 -0.3637225 -0.3657952 -0.3965186 -0.2526031 -0.2390150
#>  [4,] -0.3420530 -0.2684298 -0.2509981 -0.3399097 -0.4353340 -0.3533320
#>  [5,] -0.3118697 -0.3843941 -0.3958180 -0.2150347 -0.3074339 -0.3087914
#>  [6,] -0.2432814 -0.2288496 -0.2314416 -0.2210417 -0.3630747 -0.3841197
#>  [7,] -0.3630932 -0.3060437 -0.2472816 -0.3697651 -0.3663260 -0.2898296
#>  [8,] -0.2338590 -0.3176318 -0.3756838 -0.3415268 -0.3986573 -0.3032781
#>  [9,] -0.3546981 -0.2836392 -0.3899050 -0.4277148 -0.3272280 -0.3440821
#> [10,] -0.3024057 -0.3191598 -0.3158655 -0.3062001 -0.3039741 -0.3746976
#>              V58        V59         V6        V60          V7          V8
#>  [1,] -0.3360024 -0.3781996 -0.1603008 -0.3184624 -0.07489823 -0.20746234
#>  [2,] -0.2565629 -0.2485315 -0.1466484 -0.3297340 -0.17304387 -0.25911336
#>  [3,] -0.3225138 -0.3132134 -0.2853623 -0.4068046 -0.18800223 -0.26939892
#>  [4,] -0.4241098 -0.3181483 -0.1531286 -0.3966027 -0.15654901 -0.23279829
#>  [5,] -0.2194203 -0.3846637 -0.2723411 -0.3052838 -0.15927846 -0.16348310
#>  [6,] -0.2748036 -0.3251480 -0.1986937 -0.2842431 -0.20523278 -0.09173282
#>  [7,] -0.3850847 -0.2745607 -0.2107267 -0.2601673 -0.14606099 -0.13263310
#>  [8,] -0.2915061 -0.2965388 -0.1542361 -0.3039445 -0.23576658 -0.15559655
#>  [9,] -0.3078647 -0.3809822 -0.1201149 -0.3219694 -0.20144373 -0.09669146
#> [10,] -0.4223360 -0.3925784 -0.1032538 -0.3567871 -0.17040468 -0.12140059
#>                V9
#>  [1,] -0.10672742
#>  [2,] -0.18917089
#>  [3,] -0.08638871
#>  [4,] -0.14506811
#>  [5,] -0.10528721
#>  [6,] -0.17258261
#>  [7,] -0.20286755
#>  [8,] -0.23886925
#>  [9,] -0.04958482
#> [10,] -0.12974834
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]       [,4]        [,5]       [,6]
#> [1,] -0.01802411 -0.07051260 -0.05709537 0.02964266 -0.06147010 -0.1831334
#> [2,]  0.10488950  0.06379322 -0.01819038 0.06373588  0.07857042  0.1524991
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.14243682 -0.1042236 0.05750098 -0.03986451
#> [2,]  0.09435235  0.1094086 0.10873531  0.03173227
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 5.571066e-05 0.0004053896 0.0004732795 0.0005761215 0.0005250041
#>  [2,] 3.625781e-05 0.0002069172 0.0002103055 0.0003220881 0.0003500585
#>  [3,] 7.118715e-05 0.0004253417 0.0004693016 0.0007326518 0.0007272545
#>  [4,] 5.497590e-05 0.0003569470 0.0003789931 0.0006420399 0.0006528795
#>  [5,] 5.030851e-05 0.0003776599 0.0004528600 0.0004838601 0.0004353512
#>  [6,] 4.567106e-05 0.0005903783 0.0007835811 0.0005021388 0.0001982192
#>  [7,] 1.083435e-04 0.0007478105 0.0008951152 0.0009120660 0.0008195906
#>  [8,] 4.560354e-05 0.0004642465 0.0006008596 0.0005046624 0.0003347405
#>  [9,] 4.782421e-05 0.0003617913 0.0004155546 0.0006108287 0.0005687423
#> [10,] 3.297536e-05 0.0002465897 0.0002605660 0.0004161962 0.0003973522
#>                 V14           V15           V16           V17           V18
#>  [1,]  0.0007942748  0.0009146141  0.0010972205  0.0012475215  0.0013581930
#>  [2,]  0.0006254981  0.0007857410  0.0010064391  0.0011993778  0.0013480890
#>  [3,]  0.0014180932  0.0017775570  0.0022376172  0.0026625217  0.0029843521
#>  [4,]  0.0014238296  0.0018404003  0.0023317062  0.0027505050  0.0030391686
#>  [5,]  0.0003782967  0.0003497431  0.0004054372  0.0004584314  0.0005188244
#>  [6,] -0.0009929443 -0.0016458699 -0.0022143042 -0.0027040933 -0.0029739050
#>  [7,]  0.0007277263  0.0006723107  0.0007456859  0.0008048691  0.0008498617
#>  [8,] -0.0002495418 -0.0005692402 -0.0007938090 -0.0009828852 -0.0010576772
#>  [9,]  0.0010716344  0.0012964048  0.0016003067  0.0018584074  0.0020646447
#> [10,]  0.0008683983  0.0010888083  0.0013515938  0.0015720160  0.0017664181
#>                 V19           V2           V20           V21           V22
#>  [1,]  0.0013236988 7.130536e-05  0.0014737277  0.0015161109  0.0016971902
#>  [2,]  0.0012436460 4.799985e-05  0.0012848898  0.0013243805  0.0015307522
#>  [3,]  0.0026563804 1.001281e-04  0.0027551101  0.0028270303  0.0033808740
#>  [4,]  0.0026411218 7.868298e-05  0.0027265471  0.0028015811  0.0033374902
#>  [5,]  0.0006725601 6.556760e-05  0.0007714359  0.0008023171  0.0007941125
#>  [6,] -0.0020745424 5.462623e-05 -0.0019212028 -0.0020532393 -0.0029304782
#>  [7,]  0.0011667703 1.399339e-04  0.0014099713  0.0014938696  0.0014352408
#>  [8,] -0.0005629536 5.796634e-05 -0.0004547625 -0.0005071269 -0.0008926810
#>  [9,]  0.0018682079 6.439252e-05  0.0020019900  0.0020417618  0.0024094104
#> [10,]  0.0015569143 4.082383e-05  0.0016576438  0.0017195964  0.0020442265
#>                 V23           V24           V25           V26           V27
#>  [1,]  0.0018355126  0.0019121063  0.0019430540  0.0019075992  0.0016311487
#>  [2,]  0.0017183955  0.0018233310  0.0018755341  0.0018204873  0.0015380876
#>  [3,]  0.0038042279  0.0040387301  0.0041568806  0.0039870543  0.0032086107
#>  [4,]  0.0037511527  0.0039971380  0.0041575798  0.0040126014  0.0031716553
#>  [5,]  0.0008013979  0.0007689601  0.0006869443  0.0006751451  0.0007673620
#>  [6,] -0.0035620326 -0.0040919172 -0.0046061099 -0.0044804017 -0.0030858249
#>  [7,]  0.0013971403  0.0013207893  0.0011729210  0.0011626216  0.0013239600
#>  [8,] -0.0011603037 -0.0014043606 -0.0016798569 -0.0016258359 -0.0009478417
#>  [9,]  0.0026787900  0.0028322423  0.0029467064  0.0029035100  0.0023503246
#> [10,]  0.0022897406  0.0023948910  0.0024923377  0.0024404863  0.0019283146
#>                 V28           V29            V3           V30           V31
#>  [1,]  0.0014856142  0.0014999294  1.117629e-04  0.0014243061  0.0014504780
#>  [2,]  0.0013799637  0.0014030368  9.323091e-05  0.0014200891  0.0015427283
#>  [3,]  0.0028858261  0.0030697812  2.034010e-04  0.0031210968  0.0034758556
#>  [4,]  0.0027395409  0.0028971652  1.923054e-04  0.0029901777  0.0034514327
#>  [5,]  0.0008134844  0.0007296427  6.643445e-05  0.0005494276  0.0003204027
#>  [6,] -0.0023314532 -0.0026987707 -1.018501e-04 -0.0033190570 -0.0044540253
#>  [7,]  0.0014108684  0.0012458198  1.366990e-04  0.0009518633  0.0005178096
#>  [8,] -0.0005821353 -0.0007831391 -1.176463e-05 -0.0011698067 -0.0017918922
#>  [9,]  0.0020618177  0.0021756060  1.355997e-04  0.0021760088  0.0024437221
#> [10,]  0.0016323691  0.0017052142  1.091497e-04  0.0017501084  0.0020099879
#>                 V32           V33           V34           V35           V36
#>  [1,]  0.0013813097  0.0013450804  0.0013245940  1.209216e-03  1.162630e-03
#>  [2,]  0.0014094497  0.0014077488  0.0015106072  1.490914e-03  1.410276e-03
#>  [3,]  0.0032187984  0.0032233068  0.0034616562  3.431278e-03  3.297625e-03
#>  [4,]  0.0031741279  0.0031398306  0.0034002966  3.403555e-03  3.356658e-03
#>  [5,]  0.0003207640  0.0002926245  0.0001508538 -2.521161e-06 -5.946049e-05
#>  [6,] -0.0039707378 -0.0040973976 -0.0049500252 -5.395231e-03 -5.296387e-03
#>  [7,]  0.0005637283  0.0005168369  0.0002279330 -8.610160e-05 -2.261186e-04
#>  [8,] -0.0015780333 -0.0016579867 -0.0021145189 -2.392946e-03 -2.361337e-03
#>  [9,]  0.0023022051  0.0022656438  0.0023543308  2.299403e-03  2.277360e-03
#> [10,]  0.0018592028  0.0018281021  0.0019216612  1.890684e-03  1.872850e-03
#>                 V37           V38           V39           V4           V40
#>  [1,]  1.127494e-03  0.0010073929  0.0009149722 9.854598e-05  0.0009310828
#>  [2,]  1.287894e-03  0.0010420945  0.0008246528 8.057065e-05  0.0009129505
#>  [3,]  3.011382e-03  0.0024269068  0.0019441623 1.655205e-04  0.0021413086
#>  [4,]  3.120412e-03  0.0024414202  0.0019253765 1.308392e-04  0.0021691204
#>  [5,]  1.738086e-05  0.0001913054  0.0003150834 8.738720e-05  0.0002377282
#>  [6,] -4.497277e-03 -0.0030285896 -0.0018981369 8.152740e-06 -0.0024527822
#>  [7,] -9.594681e-05  0.0002729238  0.0005347455 1.853332e-04  0.0003496325
#>  [8,] -1.958043e-03 -0.0012066110 -0.0006304963 4.634060e-05 -0.0009290388
#>  [9,]  2.152306e-03  0.0017814791  0.0014942015 9.022140e-05  0.0015905261
#> [10,]  1.751397e-03  0.0014018886  0.0011523598 6.363321e-05  0.0013024711
#>                 V41           V42          V43          V44          V45
#>  [1,]  0.0008103476  0.0006084371 0.0005114539 0.0004778329 0.0004095582
#>  [2,]  0.0007388140  0.0004286692 0.0003187531 0.0002720845 0.0001534765
#>  [3,]  0.0016855801  0.0008982424 0.0006635883 0.0006000893 0.0002991614
#>  [4,]  0.0016908354  0.0008253458 0.0005665863 0.0005568709 0.0002272272
#>  [5,]  0.0002762004  0.0004127068 0.0003988361 0.0003500946 0.0004060590
#>  [6,] -0.0016765256 -0.0001139352 0.0002048000 0.0001716238 0.0007927894
#>  [7,]  0.0004890350  0.0008066125 0.0007863693 0.0006959852 0.0008591021
#>  [8,] -0.0005865906  0.0001780720 0.0003179297 0.0002802080 0.0005763713
#>  [9,]  0.0012606682  0.0007212265 0.0005500793 0.0005285861 0.0003222819
#> [10,]  0.0010478040  0.0005487363 0.0004075829 0.0003891041 0.0001857224
#>                V46          V47          V48          V49           V5
#>  [1,] 0.0003556655 0.0003136762 0.0002390956 1.378307e-04 1.293968e-04
#>  [2,] 0.0001478400 0.0001664989 0.0001239354 7.269271e-05 1.072415e-04
#>  [3,] 0.0002987880 0.0003644587 0.0002736885 1.587607e-04 2.145562e-04
#>  [4,] 0.0002179338 0.0003215573 0.0002413706 1.426877e-04 1.813797e-04
#>  [5,] 0.0003418039 0.0002581810 0.0001944131 1.122692e-04 1.154813e-04
#>  [6,] 0.0006618643 0.0003047035 0.0002035125 1.267424e-04 5.633766e-06
#>  [7,] 0.0007307893 0.0005267160 0.0004040538 2.289941e-04 2.222371e-04
#>  [8,] 0.0004812269 0.0002778388 0.0001943609 1.162623e-04 6.546068e-05
#>  [9,] 0.0002846487 0.0003085607 0.0002304159 1.341695e-04 1.351974e-04
#> [10,] 0.0001662661 0.0002227857 0.0001692698 1.011381e-04 9.187420e-05
#>                 V50          V51          V52           V53           V54
#>  [1,]  4.503935e-05 3.919487e-05 2.359735e-05  2.748723e-05  2.358685e-05
#>  [2,]  3.613796e-05 1.928324e-05 1.269169e-05  2.276192e-05  1.701292e-05
#>  [3,]  7.436771e-05 4.436117e-05 2.485751e-05  5.394995e-05  3.686188e-05
#>  [4,]  6.543207e-05 3.337707e-05 1.425265e-05  4.858870e-05  3.236352e-05
#>  [5,]  3.020333e-05 3.354607e-05 2.514178e-05  1.347287e-05  1.483359e-05
#>  [6,] -2.083180e-05 4.103712e-05 3.830079e-05 -3.850423e-05 -1.255468e-05
#>  [7,]  6.226731e-05 7.263387e-05 5.493935e-05  2.760314e-05  3.170834e-05
#>  [8,]  6.674589e-06 3.667096e-05 3.021972e-05 -9.409503e-06  2.249021e-06
#>  [9,]  5.179830e-05 3.530770e-05 1.708618e-05  3.812803e-05  2.727016e-05
#> [10,]  4.027307e-05 2.207035e-05 8.657152e-06  2.866689e-05  2.106376e-05
#>                 V55           V56           V57           V58           V59
#>  [1,]  2.273597e-05  2.734227e-05  2.322954e-05  1.557625e-05  1.353916e-05
#>  [2,]  2.061938e-05  2.419685e-05  2.063224e-05  1.181178e-05  1.164796e-05
#>  [3,]  4.536079e-05  5.782081e-05  4.832587e-05  2.662803e-05  2.238251e-05
#>  [4,]  4.439688e-05  5.777916e-05  4.670845e-05  2.274515e-05  1.827258e-05
#>  [5,]  8.135493e-06  7.241620e-06  7.030237e-06  9.813729e-06  1.026853e-05
#>  [6,] -4.831724e-05 -6.376577e-05 -5.050006e-05 -7.713719e-06 -5.757760e-06
#>  [7,]  1.787468e-05  1.424772e-05  1.454977e-05  2.069635e-05  2.112839e-05
#>  [8,] -1.726653e-05 -2.471764e-05 -1.911589e-05  1.938127e-06  2.690369e-06
#>  [9,]  3.295499e-05  4.214627e-05  3.383694e-05  1.882645e-05  1.454532e-05
#> [10,]  2.572789e-05  3.562800e-05  2.802019e-05  1.372796e-05  9.745372e-06
#>                  V6           V60            V7            V8           V9
#>  [1,]  0.0002931003  1.724539e-05  0.0003510465  3.907200e-04 0.0003706829
#>  [2,]  0.0002381654  1.646940e-05  0.0003051642  2.652597e-04 0.0001893922
#>  [3,]  0.0005596212  3.725112e-05  0.0007076503  6.432004e-04 0.0004004246
#>  [4,]  0.0005554923  3.375465e-05  0.0007055177  6.396049e-04 0.0003523688
#>  [5,]  0.0001336136  7.500250e-06  0.0001395903  1.849548e-04 0.0003119271
#>  [6,] -0.0004105507 -3.334428e-05 -0.0006155196 -3.180514e-04 0.0004155991
#>  [7,]  0.0002435177  1.533751e-05  0.0002453086  3.530495e-04 0.0006284013
#>  [8,] -0.0001042667 -1.079171e-05 -0.0001909173 -3.854191e-05 0.0003452823
#>  [9,]  0.0004183461  2.422846e-05  0.0005212191  5.174024e-04 0.0003396246
#> [10,]  0.0003256728  1.930760e-05  0.0004174752  4.133077e-04 0.0002445607
#> 
#> $vW[[2]]
#>            [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,]  0.1161127  0.08782681  0.08828759  0.1010291  0.06637598  0.06363565
#> [2,] -0.1161127 -0.08782681 -0.08828759 -0.1010291 -0.06637598 -0.06363565
#>            [,7]        [,8]       [,9]      [,10]
#> [1,]  0.1110774  0.05797607  0.1116302  0.1178953
#> [2,] -0.1110774 -0.05797607 -0.1116302 -0.1178953
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.6282553 0.3918287 0.3512925 0.4874553 0.4001838 0.2992881 0.4887595
#>  [8] 0.1691718 0.6362662 0.6952903
#> 
#> $B[[2]]
#> [1] 0.04248727 0.07095632
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1]  0.0026868818  0.0026097857  0.0057490523  0.0057393289  0.0009413635
#>  [6] -0.0062420362  0.0016444453 -0.0022675717  0.0041423416  0.0034371321
#> 
#> $vB[[2]]
#> [1]  0.129325 -0.129325
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0587 0.2352 0.3208 0.4257 0.5201 0.4914 0.5950 0.7221 0.9039 0.9111
#>  [2,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#>  [3,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#>  [4,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#>  [5,] 0.0108 0.0393 0.1106 0.1412 0.2202 0.2976 0.4116 0.4754 0.5390 0.6279
#>  [6,] 0.0453 0.2872 0.4918 0.6552 0.6919 0.7797 0.7464 0.9444 1.0000 0.8874
#>  [7,] 0.0473 0.3759 0.3021 0.2909 0.2301 0.1411 0.1582 0.2430 0.4474 0.5964
#>  [8,] 0.0015 0.1476 0.2118 0.2575 0.2354 0.1334 0.0092 0.1951 0.3685 0.4646
#>  [9,] 0.0091 0.0837 0.1579 0.0898 0.0309 0.1856 0.2969 0.2032 0.1264 0.1655
#> [10,] 0.0249 0.2531 0.2855 0.2961 0.3341 0.4287 0.5205 0.6087 0.7236 0.7577
#> [11,] 0.0261 0.2013 0.2890 0.3650 0.3510 0.3495 0.4325 0.5398 0.6237 0.6876
#> [12,] 0.0134 0.1250 0.2405 0.2325 0.2523 0.1472 0.0669 0.1100 0.2353 0.3282
#> [13,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#> [14,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [15,] 0.0039 0.0452 0.0492 0.0996 0.1424 0.1194 0.0628 0.0907 0.1177 0.1429
#> [16,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [17,] 0.0258 0.2954 0.2506 0.2601 0.2249 0.2115 0.1270 0.1193 0.1794 0.2185
#> [18,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [19,] 0.0409 0.2259 0.2373 0.3323 0.3827 0.4840 0.6812 0.7555 0.9522 0.9826
#> [20,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [21,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#> [22,] 0.0260 0.2354 0.2720 0.2442 0.1665 0.0336 0.1302 0.1708 0.2177 0.3175
#> [23,] 0.0353 0.0880 0.1078 0.0979 0.2250 0.2819 0.2099 0.1240 0.1699 0.0939
#> [24,] 0.0363 0.1444 0.1989 0.2154 0.2863 0.3570 0.3980 0.4359 0.5334 0.6304
#> [25,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [26,] 0.0201 0.2720 0.2188 0.3037 0.2959 0.2059 0.0906 0.1610 0.1800 0.2180
#> [27,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [28,] 0.0516 0.1294 0.2646 0.2778 0.4432 0.3672 0.2035 0.2764 0.3252 0.1536
#> [29,] 0.0269 0.5915 0.6675 0.5679 0.5175 0.3334 0.2002 0.2856 0.2937 0.3424
#> [30,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [31,] 0.0317 0.3513 0.1786 0.0658 0.0513 0.3752 0.5419 0.5440 0.5150 0.4262
#> [32,] 0.0217 0.1777 0.4079 0.4113 0.3973 0.5078 0.6509 0.8073 0.9819 1.0000
#> [33,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [34,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [35,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [36,] 0.0216 0.0973 0.1203 0.1102 0.1192 0.1762 0.2390 0.2138 0.1929 0.1765
#> [37,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [38,] 0.0096 0.2952 0.4025 0.5148 0.4901 0.4127 0.3575 0.3447 0.3068 0.2945
#> [39,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [40,] 0.0336 0.0842 0.0357 0.0689 0.1705 0.3257 0.4602 0.6225 0.7327 0.7843
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.8723 0.1210 0.7686 0.7326 0.5222 0.3097 0.3172 0.2270 0.1640 0.1746
#>  [2,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#>  [3,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#>  [4,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#>  [5,] 0.7060 0.0086 0.7918 0.9493 1.0000 0.9645 0.9432 0.8658 0.7895 0.6501
#>  [6,] 0.8024 0.0523 0.7818 0.5212 0.4052 0.3957 0.3914 0.3250 0.3200 0.3271
#>  [7,] 0.6744 0.0509 0.7969 0.8319 0.7813 0.8626 0.7369 0.4122 0.2596 0.3392
#>  [8,] 0.5418 0.0186 0.6260 0.7420 0.8257 0.8609 0.8400 0.8949 0.9945 1.0000
#>  [9,] 0.1661 0.0213 0.2091 0.2310 0.4460 0.6634 0.6933 0.7663 0.8206 0.7049
#> [10,] 0.7726 0.0119 0.8098 0.8995 0.9247 0.9365 0.9853 0.9776 1.0000 0.9896
#> [11,] 0.7329 0.0266 0.8107 0.8396 0.8632 0.8747 0.9607 0.9716 0.9121 0.8576
#> [12,] 0.4416 0.0172 0.5167 0.6508 0.7793 0.7978 0.7786 0.8587 0.9321 0.9454
#> [13,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#> [14,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [15,] 0.1223 0.0063 0.1104 0.1847 0.3715 0.4382 0.5707 0.6654 0.7476 0.7654
#> [16,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [17,] 0.1646 0.0433 0.0740 0.0625 0.2381 0.4824 0.6372 0.7531 0.8959 0.9941
#> [18,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [19,] 0.8871 0.0421 0.8268 0.7561 0.8217 0.6967 0.6444 0.6948 0.8014 0.6053
#> [20,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [21,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#> [22,] 0.3714 0.0363 0.4552 0.5700 0.7397 0.8062 0.8837 0.9432 1.0000 0.9375
#> [23,] 0.1091 0.0713 0.1410 0.1268 0.3151 0.1430 0.2264 0.5756 0.7876 0.7158
#> [24,] 0.6995 0.0478 0.7435 0.8379 0.8641 0.9014 0.9432 0.9536 1.0000 0.9547
#> [25,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [26,] 0.2026 0.0178 0.1506 0.0521 0.2143 0.4333 0.5943 0.6926 0.7576 0.8787
#> [27,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [28,] 0.2784 0.0944 0.3508 0.5187 0.7052 0.7143 0.6814 0.5100 0.5308 0.6131
#> [29,] 0.5949 0.0383 0.7526 0.8959 0.8147 0.7109 0.7378 0.7201 0.8254 0.8917
#> [30,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [31,] 0.2024 0.0956 0.4233 0.7723 0.9735 0.9390 0.5559 0.5268 0.6826 0.5713
#> [32,] 0.9407 0.0340 0.8452 0.8106 0.8460 0.6212 0.5815 0.7745 0.8204 0.5601
#> [33,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [34,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [35,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [36,] 0.0746 0.0215 0.1265 0.2005 0.1571 0.2605 0.5386 0.8440 1.0000 0.8684
#> [37,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [38,] 0.4351 0.0404 0.7264 0.8147 0.8103 0.6665 0.6958 0.7748 0.8688 1.0000
#> [39,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [40,] 0.7988 0.0294 0.8261 1.0000 0.9814 0.9620 0.9601 0.9118 0.9086 0.7931
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.1835 0.2048 0.1268 0.1674 0.2767 0.3104 0.3399 0.4441 0.5046 0.2814
#>  [2,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#>  [3,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#>  [4,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#>  [5,] 0.4492 0.4739 0.0058 0.6153 0.4929 0.3195 0.3735 0.3336 0.1052 0.0671
#>  [6,] 0.2767 0.4423 0.0843 0.2028 0.3788 0.2947 0.1984 0.2341 0.1306 0.4182
#>  [7,] 0.3788 0.4488 0.0819 0.6281 0.7449 0.7328 0.7704 0.7870 0.6048 0.5860
#>  [8,] 0.9649 0.8747 0.0289 0.6257 0.2184 0.2945 0.3645 0.5012 0.7843 0.9361
#>  [9,] 0.7560 0.7466 0.0206 0.6387 0.4846 0.3328 0.5356 0.8741 0.8573 0.6718
#> [10,] 0.9076 0.7306 0.0277 0.5758 0.4469 0.3719 0.2079 0.0955 0.0488 0.1406
#> [11,] 0.8798 0.7720 0.0223 0.5711 0.4264 0.2860 0.3114 0.2066 0.1165 0.0185
#> [12,] 0.8645 0.7220 0.0178 0.4850 0.1357 0.2951 0.4715 0.6036 0.8083 0.9870
#> [13,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#> [14,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [15,] 0.8555 0.9720 0.0152 0.9221 0.7502 0.7209 0.7757 0.6055 0.5021 0.4499
#> [16,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [17,] 0.9957 0.9328 0.0547 0.9344 0.8854 0.7690 0.6865 0.6390 0.6378 0.6629
#> [18,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [19,] 0.6084 0.8877 0.0573 0.8557 0.5563 0.2897 0.3638 0.4786 0.2908 0.0899
#> [20,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [21,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#> [22,] 0.7603 0.7123 0.0136 0.8358 0.7622 0.4567 0.1715 0.1549 0.1641 0.1869
#> [23,] 0.5998 0.5583 0.0326 0.6295 0.7659 0.8940 0.8436 0.6807 0.8380 1.0000
#> [24,] 0.9745 0.8962 0.0298 0.7196 0.5462 0.3156 0.2525 0.1969 0.2189 0.1533
#> [25,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [26,] 0.9060 0.8528 0.0274 0.9087 0.9657 0.9306 0.7774 0.6643 0.6604 0.6884
#> [27,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [28,] 0.8388 0.9031 0.0622 0.8607 0.9656 0.9168 0.7132 0.6898 0.7310 0.4134
#> [29,] 0.9820 0.8179 0.0505 0.4848 0.3203 0.2775 0.2382 0.2911 0.1675 0.3156
#> [30,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [31,] 0.5429 0.2177 0.1321 0.2149 0.5811 0.6323 0.2965 0.1873 0.2969 0.5163
#> [32,] 0.2989 0.5009 0.0392 0.6628 0.5753 0.4055 0.3746 0.3481 0.1580 0.1422
#> [33,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [34,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [35,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [36,] 0.6742 0.5537 0.0273 0.4638 0.3609 0.2055 0.1620 0.2092 0.3100 0.2344
#> [37,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [38,] 0.9941 0.8793 0.0682 0.6482 0.5876 0.6408 0.4972 0.2755 0.0300 0.3356
#> [39,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [40,] 0.5877 0.3474 0.0476 0.4235 0.4633 0.3410 0.2849 0.2847 0.1742 0.0549
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1681 0.2633 0.3198 0.1498 0.1933 0.0934 0.0443 0.0780 0.0722 0.0405
#>  [2,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#>  [3,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#>  [4,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#>  [5,] 0.0379 0.0461 0.1694 0.0460 0.2169 0.1677 0.0644 0.0159 0.0778 0.0653
#>  [6,] 0.3835 0.1057 0.1840 0.0689 0.1970 0.1674 0.0583 0.1401 0.1628 0.0621
#>  [7,] 0.6385 0.7279 0.6286 0.1252 0.5316 0.4069 0.1791 0.1625 0.2527 0.1903
#>  [8,] 0.8195 0.6207 0.4513 0.0195 0.3004 0.2674 0.2241 0.3141 0.3693 0.2986
#>  [9,] 0.3446 0.3150 0.2702 0.0505 0.2598 0.2742 0.3594 0.4382 0.2460 0.0758
#> [10,] 0.2554 0.2054 0.1614 0.0760 0.2232 0.1773 0.2293 0.2521 0.1464 0.0673
#> [11,] 0.1302 0.2480 0.1637 0.0749 0.1103 0.2144 0.2033 0.1887 0.1370 0.1376
#> [12,] 0.8800 0.6411 0.4276 0.0363 0.2702 0.2642 0.3342 0.4335 0.4542 0.3960
#> [13,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#> [14,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [15,] 0.3947 0.4281 0.4427 0.0336 0.3749 0.1972 0.0511 0.0793 0.1269 0.1533
#> [16,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [17,] 0.5983 0.4565 0.3129 0.0681 0.4158 0.4325 0.4031 0.4201 0.4557 0.3955
#> [18,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [19,] 0.2043 0.1707 0.0407 0.0130 0.1286 0.1581 0.2191 0.1701 0.0971 0.2217
#> [20,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [21,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#> [22,] 0.2655 0.1713 0.0959 0.0272 0.0768 0.0847 0.2076 0.2505 0.1862 0.1439
#> [23,] 0.9497 0.7866 0.5647 0.0272 0.3480 0.2585 0.2304 0.2948 0.3363 0.3017
#> [24,] 0.0711 0.1498 0.1755 0.0210 0.2276 0.1322 0.1056 0.1973 0.1692 0.1881
#> [25,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [26,] 0.6938 0.5932 0.5774 0.0232 0.6223 0.5841 0.4527 0.4911 0.5762 0.5013
#> [27,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [28,] 0.1580 0.1819 0.1381 0.0415 0.2960 0.6935 0.8246 0.5351 0.4403 0.6448
#> [29,] 0.1869 0.3391 0.5993 0.0707 0.4124 0.1181 0.3651 0.4655 0.4777 0.3517
#> [30,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [31,] 0.6153 0.4283 0.5479 0.1408 0.6133 0.5017 0.2377 0.1957 0.1749 0.1304
#> [32,] 0.2130 0.1866 0.1003 0.0236 0.2396 0.2241 0.2029 0.0710 0.1606 0.1669
#> [33,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [34,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [35,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [36,] 0.1058 0.0383 0.0528 0.0139 0.1291 0.2241 0.1915 0.1587 0.0942 0.0840
#> [37,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [38,] 0.3167 0.4133 0.6281 0.0688 0.4977 0.2613 0.4697 0.4806 0.4921 0.5294
#> [39,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [40,] 0.1192 0.1154 0.0855 0.0539 0.1811 0.1264 0.0799 0.0378 0.1268 0.1125
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0553 0.1081 0.1139 0.0767 0.1436 0.0265 0.0215 0.0331 0.0111 0.0088
#>  [2,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#>  [3,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#>  [4,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#>  [5,] 0.0210 0.0509 0.0387 0.0262 0.0752 0.0101 0.0161 0.0029 0.0078 0.0114
#>  [6,] 0.0203 0.0530 0.0742 0.0409 0.1183 0.0061 0.0125 0.0084 0.0089 0.0048
#>  [7,] 0.1643 0.0604 0.0209 0.0436 0.1783 0.0175 0.0107 0.0193 0.0118 0.0064
#>  [8,] 0.2226 0.0849 0.0359 0.0289 0.0515 0.0122 0.0045 0.0108 0.0075 0.0089
#>  [9,] 0.0187 0.0797 0.0748 0.0367 0.0657 0.0155 0.0300 0.0112 0.0112 0.0102
#> [10,] 0.0965 0.1492 0.1128 0.0463 0.1218 0.0193 0.0140 0.0027 0.0068 0.0150
#> [11,] 0.0307 0.0373 0.0606 0.0399 0.1364 0.0169 0.0135 0.0222 0.0175 0.0127
#> [12,] 0.2525 0.1084 0.0372 0.0286 0.0444 0.0099 0.0046 0.0094 0.0048 0.0047
#> [13,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#> [14,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [15,] 0.0690 0.0402 0.0534 0.0228 0.0310 0.0073 0.0062 0.0062 0.0120 0.0052
#> [16,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [17,] 0.2966 0.2095 0.1558 0.0884 0.0784 0.0265 0.0121 0.0091 0.0062 0.0019
#> [18,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [19,] 0.2732 0.1874 0.1062 0.0665 0.0183 0.0405 0.0113 0.0028 0.0036 0.0105
#> [20,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [21,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#> [22,] 0.1470 0.0991 0.0041 0.0154 0.0214 0.0116 0.0181 0.0146 0.0129 0.0047
#> [23,] 0.2193 0.1316 0.1078 0.0559 0.0370 0.0035 0.0098 0.0163 0.0242 0.0043
#> [24,] 0.1177 0.0779 0.0495 0.0492 0.1409 0.0194 0.0250 0.0115 0.0190 0.0055
#> [25,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [26,] 0.4042 0.3123 0.2232 0.1085 0.0724 0.0414 0.0253 0.0131 0.0049 0.0104
#> [27,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [28,] 0.6214 0.3016 0.1379 0.0364 0.0995 0.0355 0.0456 0.0432 0.0274 0.0152
#> [29,] 0.0920 0.1227 0.1785 0.1085 0.1313 0.0300 0.0346 0.0167 0.0199 0.0145
#> [30,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [31,] 0.0597 0.1124 0.1047 0.0507 0.1674 0.0159 0.0195 0.0201 0.0248 0.0131
#> [32,] 0.1700 0.1829 0.1403 0.0506 0.1081 0.0224 0.0095 0.0031 0.0103 0.0078
#> [33,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [34,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [35,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [36,] 0.0670 0.0342 0.0469 0.0357 0.0357 0.0136 0.0082 0.0140 0.0044 0.0052
#> [37,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [38,] 0.2216 0.1401 0.1888 0.0947 0.0887 0.0134 0.0310 0.0237 0.0078 0.0144
#> [39,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [40,] 0.0505 0.0949 0.0677 0.0259 0.0794 0.0170 0.0033 0.0150 0.0111 0.0032
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0158 0.0122 0.0038 0.0101 0.0228 0.0561 0.0124 0.0832 0.0672 0.1372
#>  [2,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#>  [3,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#>  [4,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#>  [5,] 0.0083 0.0058 0.0003 0.0023 0.0026 0.0887 0.0027 0.1015 0.0494 0.0472
#>  [6,] 0.0094 0.0191 0.0140 0.0049 0.0052 0.2583 0.0044 0.2156 0.3481 0.3337
#>  [7,] 0.0042 0.0054 0.0049 0.0082 0.0028 0.3070 0.0027 0.3008 0.2362 0.3830
#>  [8,] 0.0036 0.0029 0.0013 0.0010 0.0032 0.0817 0.0047 0.1005 0.0124 0.1168
#>  [9,] 0.0026 0.0097 0.0098 0.0043 0.0071 0.0795 0.0108 0.0970 0.0872 0.0743
#> [10,] 0.0012 0.0133 0.0048 0.0244 0.0077 0.1538 0.0074 0.1192 0.1229 0.2119
#> [11,] 0.0022 0.0124 0.0054 0.0021 0.0028 0.1513 0.0023 0.1316 0.1654 0.1864
#> [12,] 0.0016 0.0008 0.0042 0.0024 0.0027 0.0744 0.0041 0.0800 0.0456 0.0368
#> [13,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#> [14,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [15,] 0.0056 0.0093 0.0042 0.0003 0.0053 0.0284 0.0036 0.0396 0.0272 0.0323
#> [16,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [17,] 0.0045 0.0079 0.0031 0.0063 0.0048 0.1250 0.0050 0.1296 0.1729 0.2794
#> [18,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [19,] 0.0120 0.0087 0.0061 0.0061 0.0030 0.1019 0.0078 0.1054 0.1070 0.2302
#> [20,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [21,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#> [22,] 0.0039 0.0061 0.0040 0.0036 0.0061 0.0338 0.0115 0.0655 0.1400 0.1843
#> [23,] 0.0202 0.0108 0.0037 0.0096 0.0093 0.0792 0.0053 0.1083 0.0687 0.0298
#> [24,] 0.0096 0.0050 0.0066 0.0114 0.0073 0.1916 0.0033 0.1349 0.1613 0.1703
#> [25,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [26,] 0.0102 0.0092 0.0083 0.0020 0.0048 0.0833 0.0036 0.1232 0.1298 0.2085
#> [27,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [28,] 0.0120 0.0129 0.0020 0.0109 0.0074 0.2431 0.0078 0.1777 0.2018 0.2611
#> [29,] 0.0081 0.0045 0.0043 0.0027 0.0055 0.2103 0.0057 0.2263 0.2524 0.3595
#> [30,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [31,] 0.0070 0.0138 0.0092 0.0143 0.0036 0.1710 0.0103 0.0731 0.1401 0.2083
#> [32,] 0.0077 0.0094 0.0031 0.0030 0.0013 0.1164 0.0069 0.1398 0.1009 0.1147
#> [33,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [34,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [35,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [36,] 0.0073 0.0021 0.0047 0.0024 0.0009 0.0785 0.0017 0.0906 0.0908 0.1151
#> [37,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [38,] 0.0170 0.0012 0.0109 0.0036 0.0043 0.0932 0.0018 0.0955 0.2140 0.2546
#> [39,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [40,] 0.0035 0.0169 0.0137 0.0015 0.0069 0.0804 0.0051 0.1136 0.1228 0.1235
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7773432 0.4233939 0.5227045 0.6093282 0.3416268 0.3318858 0.7094081
#>  [2,] 0.7409759 0.3628223 0.4583955 0.5655754 0.3311187 0.2807166 0.6630685
#>  [3,] 0.9105832 0.6173092 0.7800344 0.8002403 0.5212386 0.4264488 0.8355161
#>  [4,] 0.9005408 0.5343914 0.6794010 0.7830055 0.4627003 0.3789158 0.8372590
#>  [5,] 0.9205531 0.6085882 0.7906070 0.8048198 0.5117863 0.4120146 0.8448277
#>  [6,] 0.7907783 0.3870178 0.5280235 0.5924654 0.3057414 0.3160499 0.6996103
#>  [7,] 0.8585787 0.4637626 0.4961007 0.6987839 0.3565439 0.2806428 0.7814347
#>  [8,] 0.9229310 0.5849296 0.7269081 0.8283512 0.4528196 0.3522895 0.8734684
#>  [9,] 0.8789586 0.5813521 0.6193963 0.7660574 0.4611254 0.4305610 0.8181494
#> [10,] 0.9268396 0.5409533 0.7881658 0.7972336 0.4270787 0.3291786 0.8464989
#> [11,] 0.9200225 0.5539169 0.7765500 0.7954355 0.4448438 0.3446815 0.8367543
#> [12,] 0.9060877 0.5615845 0.6696059 0.8074547 0.4438535 0.3517870 0.8598715
#> [13,] 0.9039183 0.6055113 0.6461131 0.8000300 0.4749061 0.4179042 0.8423202
#> [14,] 0.8824225 0.5168752 0.7198600 0.7555504 0.4650865 0.3749060 0.8041297
#> [15,] 0.8861253 0.6571199 0.6575999 0.7829847 0.5006418 0.4681113 0.8274955
#> [16,] 0.8698723 0.5388774 0.6589714 0.7453124 0.4356779 0.3950420 0.7908963
#> [17,] 0.8660308 0.5115928 0.5245151 0.7317706 0.3799734 0.3416050 0.7899672
#> [18,] 0.7877503 0.3803939 0.5522824 0.6068835 0.3220785 0.2924820 0.6944502
#> [19,] 0.8922000 0.5163003 0.6983163 0.7604537 0.3802868 0.3417618 0.8158498
#> [20,] 0.8677245 0.4729033 0.6706954 0.7119426 0.3908106 0.3279250 0.7955229
#> [21,] 0.9037876 0.4972147 0.7288035 0.7565862 0.4261221 0.3239746 0.8037223
#> [22,] 0.9142694 0.6144065 0.7642309 0.8110126 0.5158579 0.4064842 0.8281323
#> [23,] 0.8365802 0.5453898 0.4706995 0.6963958 0.3891946 0.3902558 0.7987822
#> [24,] 0.9267063 0.5700146 0.7823769 0.8095369 0.4486338 0.3587089 0.8434387
#> [25,] 0.9162857 0.5388262 0.7748277 0.7813701 0.4398550 0.3401791 0.8350753
#> [26,] 0.8549942 0.4724324 0.4565549 0.6992405 0.3486152 0.3085816 0.7831812
#> [27,] 0.8156336 0.4478662 0.4731342 0.6493549 0.4170455 0.3188653 0.7567451
#> [28,] 0.8499630 0.4125205 0.4392303 0.6783583 0.3117563 0.2880454 0.7398810
#> [29,] 0.8823550 0.4161450 0.6265999 0.7378154 0.3452278 0.2543574 0.7825436
#> [30,] 0.9158898 0.5850049 0.6306737 0.7849172 0.3730636 0.3487238 0.8711338
#> [31,] 0.8646369 0.4523549 0.6249560 0.6766895 0.3905616 0.3030860 0.7802433
#> [32,] 0.8879794 0.4877441 0.6801143 0.7353753 0.3808178 0.3328102 0.8042441
#> [33,] 0.9318695 0.6749514 0.7465524 0.8366800 0.5022473 0.4268713 0.8786190
#> [34,] 0.8475540 0.5151420 0.5362662 0.6814388 0.3467675 0.3488974 0.8049210
#> [35,] 0.8947874 0.5165721 0.6324250 0.7585299 0.4218640 0.3515956 0.8123154
#> [36,] 0.8570566 0.5957885 0.7086294 0.7419447 0.5181977 0.4706666 0.7790128
#> [37,] 0.9319046 0.6074740 0.8042063 0.8069875 0.4653866 0.3767214 0.8598315
#> [38,] 0.8980366 0.4544818 0.6260443 0.7485317 0.3577708 0.2787589 0.8101899
#> [39,] 0.9104101 0.5780575 0.7615667 0.8027598 0.4683661 0.3662791 0.8272329
#> [40,] 0.9214978 0.5850557 0.7977593 0.7964203 0.4820871 0.3795873 0.8451147
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3018120 0.7509950 0.7779359
#>  [2,] 0.2788216 0.6952682 0.7316427
#>  [3,] 0.4592653 0.8849249 0.8797733
#>  [4,] 0.3771949 0.8784809 0.8698950
#>  [5,] 0.4240946 0.8997426 0.8920967
#>  [6,] 0.2944539 0.7563869 0.7877885
#>  [7,] 0.2528662 0.8149267 0.8513624
#>  [8,] 0.3851320 0.9085944 0.9120994
#>  [9,] 0.4190586 0.8530200 0.8745861
#> [10,] 0.3714693 0.9062672 0.8957462
#> [11,] 0.3814079 0.9013964 0.8924341
#> [12,] 0.3804049 0.8915333 0.8945616
#> [13,] 0.3990903 0.8823982 0.8941683
#> [14,] 0.3877027 0.8693095 0.8624499
#> [15,] 0.4352116 0.8512750 0.8848410
#> [16,] 0.3619631 0.8558493 0.8593751
#> [17,] 0.3366714 0.8240867 0.8585676
#> [18,] 0.2893818 0.7673712 0.7668577
#> [19,] 0.3397582 0.8757434 0.8860109
#> [20,] 0.3521439 0.8593904 0.8581231
#> [21,] 0.3663043 0.8772475 0.8679780
#> [22,] 0.4520208 0.8843879 0.8832768
#> [23,] 0.3407019 0.7755921 0.8444437
#> [24,] 0.3909180 0.9076188 0.8994273
#> [25,] 0.3770988 0.8958166 0.8863511
#> [26,] 0.2897708 0.7981855 0.8497968
#> [27,] 0.2748432 0.7476017 0.8071279
#> [28,] 0.2716784 0.8226277 0.8239298
#> [29,] 0.2860969 0.8514673 0.8386321
#> [30,] 0.3282712 0.8835035 0.9232628
#> [31,] 0.3117749 0.8273034 0.8289723
#> [32,] 0.3275565 0.8544171 0.8688759
#> [33,] 0.4351672 0.9116477 0.9211134
#> [34,] 0.3251455 0.8195459 0.8670684
#> [35,] 0.3307171 0.8628479 0.8692826
#> [36,] 0.4850305 0.8286660 0.8371032
#> [37,] 0.4037732 0.9165846 0.9078975
#> [38,] 0.2884033 0.8626563 0.8577754
#> [39,] 0.4125347 0.8908456 0.8768330
#> [40,] 0.4037173 0.9064301 0.8914545
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6575477 0.3424523
#>  [2,] 0.6516222 0.3483778
#>  [3,] 0.6778663 0.3221337
#>  [4,] 0.6766239 0.3233761
#>  [5,] 0.6808945 0.3191055
#>  [6,] 0.6586352 0.3413648
#>  [7,] 0.6707648 0.3292352
#>  [8,] 0.6842598 0.3157402
#>  [9,] 0.6700930 0.3299070
#> [10,] 0.6854388 0.3145612
#> [11,] 0.6837806 0.3162194
#> [12,] 0.6800995 0.3199005
#> [13,] 0.6757087 0.3242913
#> [14,] 0.6756889 0.3243111
#> [15,] 0.6708830 0.3291170
#> [16,] 0.6724686 0.3275314
#> [17,] 0.6693985 0.3306015
#> [18,] 0.6606495 0.3393505
#> [19,] 0.6791663 0.3208337
#> [20,] 0.6743716 0.3256284
#> [21,] 0.6796628 0.3203372
#> [22,] 0.6790474 0.3209526
#> [23,] 0.6612580 0.3387420
#> [24,] 0.6843875 0.3156125
#> [25,] 0.6828780 0.3171220
#> [26,] 0.6667927 0.3332073
#> [27,] 0.6605610 0.3394390
#> [28,] 0.6663451 0.3336549
#> [29,] 0.6777151 0.3222849
#> [30,] 0.6806014 0.3193986
#> [31,] 0.6712087 0.3287913
#> [32,] 0.6766292 0.3233708
#> [33,] 0.6820834 0.3179166
#> [34,] 0.6675225 0.3324775
#> [35,] 0.6757834 0.3242166
#> [36,] 0.6659792 0.3340208
#> [37,] 0.6848747 0.3151253
#> [38,] 0.6786504 0.3213496
#> [39,] 0.6809989 0.3190011
#> [40,] 0.6827836 0.3172164
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>           [,1]        [,2]        [,3]      [,4]         [,5]       [,6]
#>  [1,] 1.250250 -0.30885650  0.09088067 0.4444892 -0.656052936 -0.6996683
#>  [2,] 1.051047 -0.56313535 -0.16680349 0.2638213 -0.703129809 -0.9409099
#>  [3,] 2.320777  0.47814270  1.26586712 1.3877972  0.085005460 -0.2963550
#>  [4,] 2.203248  0.13778333  0.75102031 1.2832677 -0.149476549 -0.4941526
#>  [5,] 2.449886  0.44138169  1.32858807 1.4166951  0.047153767 -0.3556434
#>  [6,] 1.329623 -0.45986541  0.11221155 0.3741667 -0.820105079 -0.7719850
#>  [7,] 1.803535 -0.14520426 -0.01559745 0.8415135 -0.590396597 -0.9412753
#>  [8,] 2.482854  0.34304348  0.97899148 1.5739867 -0.189284772 -0.6089907
#>  [9,] 1.982605  0.32832619  0.48698641 1.1861811 -0.155812737 -0.2795626
#> [10,] 2.539126  0.16418090  1.31390450 1.3690931 -0.293780124 -0.7119025
#> [11,] 2.442653  0.21650948  1.24567344 1.3580066 -0.221526165 -0.6425008
#> [12,] 2.266775  0.24759497  0.70640329 1.4335555 -0.225537342 -0.6111933
#> [13,] 2.241540  0.42848285  0.60199715 1.3864819 -0.100460015 -0.3313827
#> [14,] 2.015573  0.06752636  0.94376750 1.1284371 -0.139881542 -0.5112267
#> [15,] 2.051759  0.65048568  0.65261683 1.2831453  0.002567277 -0.1277283
#> [16,] 1.899830  0.15582415  0.65871379 1.0737659 -0.258722141 -0.4261669
#> [17,] 1.866311  0.04637950  0.09813920 1.0036246 -0.489661234 -0.6561499
#> [18,] 1.311418 -0.48787648  0.20989671 0.4342307 -0.744236290 -0.8833600
#> [19,] 2.113412  0.06522428  0.83929300 1.1551685 -0.488331280 -0.6554530
#> [20,] 1.880987 -0.10849329  0.71133178 0.9048374 -0.443905969 -0.7175849
#> [21,] 2.240036 -0.01114130  0.98856068 1.1340533 -0.297690865 -0.7355655
#> [22,] 2.366915  0.46587288  1.17601680 1.4566030  0.063452747 -0.3785187
#> [23,] 1.633000  0.18206027 -0.11733646 0.8301934 -0.450698846 -0.4462371
#> [24,] 2.537162  0.28191074  1.27957210 1.4470039 -0.206192091 -0.5809721
#> [25,] 2.392918  0.15561796  1.23577475 1.2736687 -0.241750461 -0.6624961
#> [26,] 1.774321 -0.11038233 -0.17421965 0.8436839 -0.625131648 -0.8067586
#> [27,] 1.487040 -0.20929571 -0.10756691 0.6162046 -0.334913942 -0.7589912
#> [28,] 1.734311 -0.35355561 -0.24428623 0.7462376 -0.791921211 -0.9048959
#> [29,] 2.014923 -0.33861884  0.51765758 1.0346450 -0.640082831 -1.0755061
#> [30,] 2.387768  0.34335348  0.53510799 1.2945554 -0.519096210 -0.6246536
#> [31,] 1.854349 -0.19116030  0.51063774 0.7385993 -0.444952091 -0.8326456
#> [32,] 2.070266 -0.04903334  0.75429724 1.0220686 -0.486078703 -0.6955024
#> [33,] 2.615767  0.73066611  1.08030853 1.6337301  0.008989340 -0.2946276
#> [34,] 1.715544  0.06058666  0.14532019 0.7603920 -0.633278732 -0.6238894
#> [35,] 2.140603  0.06631285  0.54263416 1.1446368 -0.315125954 -0.6120328
#> [36,] 1.791055  0.38794733  0.88873671 1.0561009  0.072822884 -0.1174684
#> [37,] 2.616321  0.43670671  1.41279420 1.4305532 -0.138675508 -0.5034875
#> [38,] 2.175597 -0.18257811  0.51528383 1.0907967 -0.585052509 -0.9506261
#> [39,] 2.318652  0.31480448  1.16128812 1.4036330 -0.126704770 -0.5482131
#> [40,] 2.462873  0.34356264  1.37234824 1.3640694 -0.071682237 -0.4913001
#>            [,7]        [,8]      [,9]    [,10]
#>  [1,] 0.8925111 -0.83868415 1.1039261 1.253678
#>  [2,] 0.6769988 -0.95031449 0.8248657 1.002973
#>  [3,] 1.6252372 -0.16330087 2.0399181 1.990285
#>  [4,] 1.6379733 -0.50147151 1.9781231 1.900030
#>  [5,] 1.6945969 -0.30598701 2.1943677 2.112339
#>  [6,] 0.8454427 -0.87384977 1.1329719 1.311646
#>  [7,] 1.2740467 -1.08338369 1.4823465 1.745327
#>  [8,] 1.9319800 -0.46782151 2.2965919 2.339542
#>  [9,] 1.5038593 -0.32663899 1.7584868 1.942131
#> [10,] 1.7074011 -0.52591865 2.2688858 2.150829
#> [11,] 1.6342741 -0.48357662 2.2128375 2.115849
#> [12,] 1.8142232 -0.48782983 2.1065000 2.138207
#> [13,] 1.6755937 -0.40925685 2.0153394 2.134044
#> [14,] 1.4123079 -0.45697907 1.8948674 1.835789
#> [15,] 1.5679801 -0.26061888 1.7446359 2.039094
#> [16,] 1.3303366 -0.56685381 1.7812353 1.810109
#> [17,] 1.3247279 -0.67816321 1.5442846 1.803444
#> [18,] 0.8210079 -0.89838816 1.1935269 1.190652
#> [19,] 1.4884788 -0.66437210 1.9527244 2.050626
#> [20,] 1.3585436 -0.60962869 1.8102357 1.799788
#> [21,] 1.4097235 -0.54810434 1.9666188 1.883198
#> [22,] 1.5724483 -0.19250914 2.0346555 2.023833
#> [23,] 1.3787007 -0.66016772 1.2401614 1.691671
#> [24,] 1.6840398 -0.44345488 2.2849004 2.190877
#> [25,] 1.6220326 -0.50188086 2.1515831 2.053999
#> [26,] 1.2843018 -0.89649740 1.3749920 1.733008
#> [27,] 1.1349164 -0.97018701 1.0858622 1.431455
#> [28,] 1.0453499 -0.98612344 1.5342527 1.543203
#> [29,] 1.2805513 -0.91441652 1.7461563 1.648085
#> [30,] 1.9110213 -0.71601443 2.0260336 2.487528
#> [31,] 1.2670846 -0.79183459 1.5666352 1.578361
#> [32,] 1.4130343 -0.71925773 1.7696741 1.891056
#> [33,] 1.9794172 -0.26079963 2.3339217 2.457571
#> [34,] 1.4173398 -0.73022445 1.5132737 1.875283
#> [35,] 1.4651260 -0.70494364 1.8391477 1.894630
#> [36,] 1.2599226 -0.05989589 1.5762021 1.636831
#> [37,] 1.8138914 -0.38976765 2.3968220 2.288229
#> [38,] 1.4512442 -0.90315126 1.8375299 1.796935
#> [39,] 1.5661415 -0.35349677 2.0994077 1.962776
#> [40,] 1.6967872 -0.38999986 2.2708046 2.105685
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.4648491 -0.1875358
#>  [2,] 0.4480520 -0.1781253
#>  [3,] 0.5231922 -0.2207912
#>  [4,] 0.5226535 -0.2156459
#>  [5,] 0.5314281 -0.2264573
#>  [6,] 0.4675441 -0.1896738
#>  [7,] 0.5071197 -0.2045266
#>  [8,] 0.5431492 -0.2302689
#>  [9,] 0.5058071 -0.2027986
#> [10,] 0.5430591 -0.2358215
#> [11,] 0.5390159 -0.2321848
#> [12,] 0.5329819 -0.2212473
#> [13,] 0.5221886 -0.2119314
#> [14,] 0.5171592 -0.2168703
#> [15,] 0.5063971 -0.2057845
#> [16,] 0.5096105 -0.2097610
#> [17,] 0.5046738 -0.2007922
#> [18,] 0.4732906 -0.1928993
#> [19,] 0.5263981 -0.2235452
#> [20,] 0.5128731 -0.2151514
#> [21,] 0.5280001 -0.2242227
#> [22,] 0.5275267 -0.2218709
#> [23,] 0.4805338 -0.1883715
#> [24,] 0.5412656 -0.2327435
#> [25,] 0.5358435 -0.2311864
#> [26,] 0.4976517 -0.1960627
#> [27,] 0.4761055 -0.1896899
#> [28,] 0.4994351 -0.1922656
#> [29,] 0.5244991 -0.2187920
#> [30,] 0.5331476 -0.2233892
#> [31,] 0.5037927 -0.2098644
#> [32,] 0.5187014 -0.2196225
#> [33,] 0.5373600 -0.2260030
#> [34,] 0.4956376 -0.2013634
#> [35,] 0.5204214 -0.2140394
#> [36,] 0.4901611 -0.1998942
#> [37,] 0.5417703 -0.2344952
#> [38,] 0.5276421 -0.2199341
#> [39,] 0.5323724 -0.2259938
#> [40,] 0.5359825 -0.2306113
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3424523 -0.3424523
#>  [2,]  0.3483778 -0.3483778
#>  [3,] -0.6778663  0.6778663
#>  [4,] -0.6766239  0.6766239
#>  [5,] -0.6808945  0.6808945
#>  [6,] -0.6586352  0.6586352
#>  [7,] -0.6707648  0.6707648
#>  [8,]  0.3157402 -0.3157402
#>  [9,] -0.6700930  0.6700930
#> [10,]  0.3145612 -0.3145612
#> [11,]  0.3162194 -0.3162194
#> [12,]  0.3199005 -0.3199005
#> [13,] -0.6757087  0.6757087
#> [14,] -0.6756889  0.6756889
#> [15,] -0.6708830  0.6708830
#> [16,]  0.3275314 -0.3275314
#> [17,]  0.3306015 -0.3306015
#> [18,]  0.3393505 -0.3393505
#> [19,] -0.6791663  0.6791663
#> [20,] -0.6743716  0.6743716
#> [21,]  0.3203372 -0.3203372
#> [22,]  0.3209526 -0.3209526
#> [23,] -0.6612580  0.6612580
#> [24,]  0.3156125 -0.3156125
#> [25,]  0.3171220 -0.3171220
#> [26,]  0.3332073 -0.3332073
#> [27,] -0.6605610  0.6605610
#> [28,]  0.3336549 -0.3336549
#> [29,]  0.3222849 -0.3222849
#> [30,] -0.6806014  0.6806014
#> [31,] -0.6712087  0.6712087
#> [32,] -0.6766292  0.6766292
#> [33,] -0.6820834  0.6820834
#> [34,] -0.6675225  0.6675225
#> [35,] -0.6757834  0.6757834
#> [36,] -0.6659792  0.6659792
#> [37,] -0.6848747  0.6848747
#> [38,]  0.3213496 -0.3213496
#> [39,]  0.3190011 -0.3190011
#> [40,] -0.6827836  0.6827836
#> 
#> $L
#> [1] 0.6927002 0.7176327 0.7860134 0.6992637 0.7118332 0.7932146
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5652174