Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12         V13          V14
#>  [1,] -0.3686177 -0.08330515 -0.09613497 -0.14066110 -0.13859658 -0.159819516
#>  [2,] -0.1729300 -0.22294867 -0.16872050 -0.01060377 -0.01064350 -0.029264020
#>  [3,] -0.1503332 -0.02803797  0.04202849  0.05641112 -0.10009447 -0.068136322
#>  [4,] -0.4007214 -0.03484986 -0.11798897 -0.12620363 -0.06444795  0.007758374
#>  [5,] -0.2922270 -0.09397110 -0.02626563 -0.12013884 -0.12955267  0.022329742
#>  [6,] -0.2864565 -0.10715241 -0.09909294 -0.09810457 -0.14552943 -0.168500897
#>  [7,] -0.3512774 -0.02222521 -0.01471639 -0.07961411 -0.07340909  0.014764359
#>  [8,] -0.3325096 -0.03644286 -0.16357569 -0.16593545 -0.11690628 -0.085392143
#>  [9,] -0.3302880 -0.24418686 -0.03902297 -0.13976426 -0.04069502 -0.109225536
#> [10,] -0.3589813 -0.09083045 -0.16944092 -0.12701250 -0.12813690 -0.046660014
#>               V15          V16         V17         V18          V19         V2
#>  [1,] -0.11569044 -0.100393967 -0.04038000  0.05420767  0.050459096 -0.2843901
#>  [2,] -0.13785283 -0.166118261 -0.10284073 -0.03668403 -0.023428447 -0.2895611
#>  [3,] -0.03079460  0.026630157  0.03997864 -0.06149824 -0.007225465 -0.2369481
#>  [4,] -0.01496632 -0.060160751  0.06992191 -0.03019240  0.150172548 -0.2745804
#>  [5,] -0.10655104 -0.084722293 -0.13427033 -0.07821835 -0.037401194 -0.1624274
#>  [6,] -0.16261965 -0.118419032 -0.06457235 -0.16926567 -0.015288005 -0.1944897
#>  [7,] -0.10550721 -0.099614921 -0.06482226 -0.07408488 -0.030852304 -0.2296291
#>  [8,]  0.06483448  0.068345514 -0.05109098  0.07199998  0.065659163 -0.3413349
#>  [9,]  0.07036870 -0.003993252 -0.06468206  0.08394717 -0.009168744 -0.3939977
#> [10,] -0.05643870  0.081155980 -0.07861973  0.03618647 -0.000197528 -0.2751985
#>                V20         V21         V22         V23         V24          V25
#>  [1,]  0.019288273 -0.05374071  0.03900465 -0.03456588 -0.04702948 -0.003979639
#>  [2,] -0.105037954 -0.05387156  0.04610162 -0.07463616 -0.03973433  0.093381613
#>  [3,]  0.189009161  0.06320343  0.05618005  0.19000458  0.25760928  0.095271457
#>  [4,]  0.168543847  0.15825499  0.24046556  0.11240243  0.26640172  0.225975192
#>  [5,] -0.018689786 -0.06858420 -0.06053064 -0.03432744  0.10100979  0.074508506
#>  [6,] -0.041863613  0.10264091  0.01627453  0.07221158  0.04745736  0.115320910
#>  [7,] -0.034844901  0.10712142  0.19754291  0.15462092  0.10245923  0.310633487
#>  [8,] -0.004214217  0.06777391  0.16435208  0.23317457  0.20369274  0.103396782
#>  [9,]  0.153184497  0.09621379  0.14367175  0.22599962  0.13045507  0.132363103
#> [10,]  0.167186831  0.13790172  0.09529647  0.28585984  0.25063480  0.120450014
#>               V26         V27        V28         V29         V3         V30
#>  [1,] -0.02501284  0.08665304 0.04627816  0.09157159 -0.2051951  0.03008033
#>  [2,]  0.15576830  0.07834069 0.08685418 -0.01844969 -0.3602892  0.05094939
#>  [3,]  0.12790118  0.28011955 0.22777268  0.11958474 -0.1481839  0.06165458
#>  [4,]  0.20231987  0.12540851 0.25524475  0.16891585 -0.2579258  0.13325379
#>  [5,]  0.13350824  0.07891488 0.22797764  0.18307898 -0.1665512  0.18495908
#>  [6,]  0.11783830 -0.03392981 0.02786454 -0.03353256 -0.2317486 -0.01765589
#>  [7,]  0.34455068  0.19310856 0.17072903  0.10293477 -0.3137994  0.11114148
#>  [8,]  0.26162532  0.05528062 0.05146051  0.05031931 -0.2219982  0.08982898
#>  [9,]  0.17644155  0.19353604 0.10090280  0.04139046 -0.3647183  0.06743306
#> [10,]  0.13634774  0.26253752 0.13496265  0.12020141 -0.3905894  0.12723667
#>                V31          V32         V33         V34          V35
#>  [1,] -0.019940442  0.003295573 -0.03600094 -0.15465300 -0.024630720
#>  [2,]  0.105975450  0.070712848 -0.02492756 -0.07677748 -0.041400286
#>  [3,]  0.087761024  0.042347901  0.13810563  0.16020595  0.127441460
#>  [4,]  0.051533985  0.013562041  0.10753988  0.11969830 -0.045174150
#>  [5,]  0.061910784 -0.041043127  0.01263655 -0.04075918  0.049002497
#>  [6,] -0.106406868 -0.052461653 -0.01435213 -0.04690299 -0.008712628
#>  [7,]  0.001291852  0.014920775  0.09305774 -0.06632631  0.013907101
#>  [8,]  0.050879255  0.167485977  0.08477402 -0.03987924  0.083513965
#>  [9,]  0.052769533 -0.072761791 -0.14000488 -0.03656820 -0.121351258
#> [10,]  0.136195234  0.037759074 -0.04594397  0.04232082 -0.029845962
#>                V36          V37           V38          V39         V4
#>  [1,] -0.049032594 -0.113032401 -0.0680857044 -0.210289267 -0.2777413
#>  [2,]  0.005593165 -0.021759729 -0.0309261954 -0.111437834 -0.2556917
#>  [3,]  0.052651224  0.132527018  0.0160442375 -0.083155384 -0.2849009
#>  [4,] -0.051019629 -0.053562758 -0.0002689388  0.050081842 -0.2425493
#>  [5,] -0.044391236 -0.073809502 -0.1491647522  0.020436808 -0.1997666
#>  [6,]  0.011042789  0.008253818 -0.1554487903  0.009499375 -0.2491519
#>  [7,]  0.086492996  0.040616165 -0.0049173317 -0.060112779 -0.2138004
#>  [8,] -0.034804150  0.038201356 -0.0642382324  0.009287354 -0.1750438
#>  [9,] -0.052054171 -0.123906065 -0.0554765810 -0.062695810 -0.2365978
#> [10,] -0.121902911 -0.040082441 -0.0198054106 -0.019506567 -0.3997428
#>                V40          V41          V42          V43         V44
#>  [1,] -0.097174967 -0.157552175 -0.070419831 -0.140888616 -0.09464117
#>  [2,] -0.172498993 -0.179159524 -0.184928118 -0.026515501 -0.05058975
#>  [3,]  0.070071379  0.041607604  0.008026214 -0.005594385 -0.09398681
#>  [4,]  0.018202500 -0.080088021 -0.088465056 -0.060103824 -0.15190176
#>  [5,] -0.026622823 -0.100550924 -0.043284725 -0.004144861 -0.10923365
#>  [6,] -0.061398051 -0.123360803 -0.204714983 -0.177123079 -0.17532717
#>  [7,] -0.010667306 -0.027506288 -0.008523118 -0.112104957 -0.16482944
#>  [8,] -0.066367383  0.021380832 -0.023348961 -0.158289819 -0.10677870
#>  [9,] -0.071160416 -0.128215339 -0.204989041 -0.043438369 -0.12193205
#> [10,]  0.008287949 -0.007925065  0.031144445  0.003085145 -0.06234694
#>                V45         V46         V47         V48        V49         V5
#>  [1,] -0.102188298 -0.21988376 -0.18644865 -0.31296497 -0.2157174 -0.2948006
#>  [2,] -0.184094471 -0.19169189 -0.17748431 -0.30575940 -0.2794390 -0.1367486
#>  [3,] -0.104886486 -0.11046641 -0.04742881 -0.15773487 -0.2972043 -0.2372319
#>  [4,] -0.004568575 -0.08997160 -0.29091576 -0.17400824 -0.3656250 -0.3030159
#>  [5,] -0.120762674 -0.07013123 -0.08062787 -0.09872664 -0.2693836 -0.2734009
#>  [6,] -0.187634585 -0.17036552 -0.13895427 -0.15517598 -0.1219333 -0.1081129
#>  [7,] -0.183631592 -0.08205161 -0.06060518 -0.12587977 -0.2505419 -0.2473178
#>  [8,] -0.175035956 -0.14120877 -0.22642652 -0.27998743 -0.1866694 -0.1022479
#>  [9,] -0.196648007 -0.25695994 -0.13715622 -0.17690342 -0.2525555 -0.2620002
#> [10,] -0.146643193 -0.14213377 -0.20085673 -0.16720605 -0.3497857 -0.2008762
#>              V50        V51        V52        V53        V54        V55
#>  [1,] -0.2672207 -0.3036685 -0.2469349 -0.2266456 -0.4176973 -0.2670150
#>  [2,] -0.3771655 -0.2555929 -0.2934489 -0.3036820 -0.3780361 -0.2892597
#>  [3,] -0.2931243 -0.3506377 -0.2723966 -0.3146646 -0.3656035 -0.3516049
#>  [4,] -0.3320404 -0.3208253 -0.3184419 -0.3637516 -0.3761160 -0.4478224
#>  [5,] -0.3393795 -0.3200593 -0.2849235 -0.2455885 -0.2171613 -0.2300247
#>  [6,] -0.2141222 -0.3083642 -0.3049909 -0.2112403 -0.2044534 -0.3254019
#>  [7,] -0.3378036 -0.3305509 -0.2518647 -0.3707748 -0.2512030 -0.3882840
#>  [8,] -0.2663159 -0.3519891 -0.4120557 -0.3241639 -0.2612878 -0.3927568
#>  [9,] -0.3503512 -0.2436471 -0.3173460 -0.2651794 -0.3149825 -0.2225305
#> [10,] -0.3620488 -0.3303965 -0.4069121 -0.4235696 -0.4043776 -0.3725725
#>              V56        V57        V58        V59         V6        V60
#>  [1,] -0.3466515 -0.2818280 -0.3859728 -0.3568200 -0.1659313 -0.3321695
#>  [2,] -0.2502682 -0.3776344 -0.2544029 -0.3265410 -0.2167056 -0.3250828
#>  [3,] -0.3160766 -0.3236517 -0.3292110 -0.2889350 -0.2253511 -0.2578604
#>  [4,] -0.3179228 -0.4484355 -0.4029648 -0.3736274 -0.1473698 -0.4896621
#>  [5,] -0.3543090 -0.3326749 -0.3764038 -0.3370193 -0.2082868 -0.2205760
#>  [6,] -0.1755456 -0.3095334 -0.2250433 -0.3197549 -0.1230680 -0.2859636
#>  [7,] -0.3377541 -0.2501434 -0.3094008 -0.3578556 -0.1458048 -0.3319377
#>  [8,] -0.3978184 -0.3185784 -0.3710897 -0.2717679 -0.1647653 -0.2816821
#>  [9,] -0.4093487 -0.3467520 -0.2811825 -0.3981408 -0.2148102 -0.2524902
#> [10,] -0.2993448 -0.3448284 -0.3617979 -0.3234088 -0.2033420 -0.3190976
#>                V7          V8          V9
#>  [1,] -0.22978543 -0.21584269 -0.23718286
#>  [2,] -0.12563671 -0.09705554 -0.20967573
#>  [3,] -0.13837195 -0.19113628 -0.12468989
#>  [4,] -0.13621412 -0.10050055 -0.13319249
#>  [5,] -0.06864485 -0.17492194 -0.14544717
#>  [6,] -0.17641533 -0.14347160 -0.04354615
#>  [7,] -0.19871837 -0.24392993 -0.19193713
#>  [8,] -0.21022484 -0.13073990 -0.15288435
#>  [9,] -0.20901564 -0.21522650 -0.17016587
#> [10,] -0.18639500 -0.13285724 -0.08263625
#> 
#> $W[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.07999395 -0.00374633  0.12774925  0.05374182  0.09044592 0.01164588
#> [2,]  0.09296001  0.07313188 -0.07203106 -0.18264449 -0.09144775 0.09050473
#>            [,7]        [,8]       [,9]       [,10]
#> [1,] 0.14633400 -0.03510167  0.1089222  0.09299688
#> [2,] 0.01654445 -0.02985465 -0.1209370 -0.02441961
#> 
#> 
#> $vW
#> $vW[[1]]
#>                  V1           V10           V11           V12           V13
#>  [1,]  6.671850e-04  4.735451e-03  6.048787e-03  6.275755e-03  0.0067945663
#>  [2,]  4.180447e-04  3.075306e-03  3.896972e-03  4.013751e-03  0.0044131674
#>  [3,]  1.323324e-04  9.893428e-04  1.232563e-03  1.241039e-03  0.0015094226
#>  [4,]  7.983046e-05  6.714000e-04  8.315140e-04  8.166355e-04  0.0010236586
#>  [5,] -2.473107e-05  5.795207e-05  1.898718e-05 -6.565057e-06  0.0002123335
#>  [6,]  3.657247e-04  2.741041e-03  3.456346e-03  3.554793e-03  0.0038977776
#>  [7,]  2.497813e-04  1.777233e-03  2.228424e-03  2.216021e-03  0.0026523481
#>  [8,]  4.906678e-04  3.424543e-03  4.284250e-03  4.257659e-03  0.0048617441
#>  [9,] -6.838107e-05 -2.171151e-04 -3.378703e-04 -3.791702e-04 -0.0001751833
#> [10,]  2.739650e-04  1.950240e-03  2.441552e-03  2.414374e-03  0.0028282088
#>                V14          V15         V16         V17         V18         V19
#>  [1,] 0.0066763865 0.0062927274 0.006924160 0.007560951 0.008585686 0.010846922
#>  [2,] 0.0044678293 0.0043579811 0.004904880 0.005375536 0.005996756 0.007359498
#>  [3,] 0.0017814069 0.0019547966 0.002298090 0.002515920 0.002588374 0.002844217
#>  [4,] 0.0012559817 0.0014154825 0.001688298 0.001858724 0.001881066 0.001999265
#>  [5,] 0.0007123878 0.0011657300 0.001647547 0.001894246 0.001816293 0.001523950
#>  [6,] 0.0039312048 0.0037990215 0.004272318 0.004665144 0.005270011 0.006517301
#>  [7,] 0.0029784254 0.0031564661 0.003637345 0.003966498 0.004147176 0.004723990
#>  [8,] 0.0050216514 0.0049958030 0.005510824 0.005896946 0.006287584 0.007620500
#>  [9,] 0.0003900236 0.0009063324 0.001393366 0.001622279 0.001486819 0.001017916
#> [10,] 0.0030512746 0.0031519432 0.003570481 0.003857800 0.004038812 0.004690458
#>                  V2          V20          V21         V22         V23
#>  [1,]  6.481100e-04 0.0125337957 0.0135165503 0.013325564 0.012531308
#>  [2,]  4.002621e-04 0.0084537308 0.0091558029 0.009155985 0.008727675
#>  [3,]  1.285702e-04 0.0031495538 0.0033605429 0.003331954 0.003148020
#>  [4,]  7.415804e-05 0.0022075056 0.0023671004 0.002375565 0.002254715
#>  [5,] -3.825174e-05 0.0014562740 0.0015961963 0.001859646 0.001975756
#>  [6,]  3.515872e-04 0.0075297175 0.0081529928 0.008108129 0.007707632
#>  [7,]  2.451130e-04 0.0053286390 0.0056725348 0.005534176 0.005180354
#>  [8,]  5.041243e-04 0.0089189406 0.0095661357 0.009159247 0.008484290
#>  [9,] -8.381349e-05 0.0008346475 0.0009289352 0.001256744 0.001452259
#> [10,]  2.748792e-04 0.0053918054 0.0057561292 0.005542200 0.005136512
#>               V24         V25         V26         V27         V28         V29
#>  [1,] 0.011699338 0.011907931 0.012309204 0.011812675 0.011715265 0.009948303
#>  [2,] 0.008234293 0.008443416 0.008737723 0.008313333 0.008170817 0.007057665
#>  [3,] 0.002917232 0.003038142 0.003195780 0.002913383 0.002812984 0.002544977
#>  [4,] 0.002101807 0.002241854 0.002367545 0.002103448 0.002005547 0.001841035
#>  [5,] 0.001986229 0.002141498 0.002210424 0.001843112 0.001548498 0.001644690
#>  [6,] 0.007261168 0.007395758 0.007643588 0.007273664 0.007159609 0.006156756
#>  [7,] 0.004770199 0.004945133 0.005197887 0.004754898 0.004655285 0.004136957
#>  [8,] 0.007844162 0.008168086 0.008670543 0.008223177 0.008282850 0.007118958
#>  [9,] 0.001523401 0.001679779 0.001721425 0.001331972 0.001039338 0.001282487
#> [10,] 0.004721099 0.004945652 0.005245235 0.004858193 0.004833905 0.004223277
#>                  V3         V30         V31         V32         V33         V34
#>  [1,]  7.089570e-04 0.007800555 0.005274442 0.004677041 0.004044795 0.002769119
#>  [2,]  4.512271e-04 0.005674319 0.004039239 0.003497904 0.003092508 0.002379995
#>  [3,]  1.602445e-04 0.002257321 0.001766931 0.001459533 0.001411796 0.001397586
#>  [4,]  1.023916e-04 0.001672669 0.001340657 0.001100297 0.001072786 0.001107640
#>  [5,] -1.827506e-06 0.001704034 0.001707815 0.001232989 0.001274404 0.001620571
#>  [6,]  3.970919e-04 0.004927808 0.003426878 0.002966015 0.002640445 0.002002543
#>  [7,]  2.957047e-04 0.003599714 0.002656915 0.002274682 0.002193080 0.002070164
#>  [8,]  5.751328e-04 0.005946305 0.004117585 0.003707727 0.003459997 0.002946814
#>  [9,] -4.807998e-05 0.001468020 0.001608254 0.001108381 0.001204588 0.001664855
#> [10,]  3.269701e-04 0.003641787 0.002613323 0.002301758 0.002205056 0.002014170
#>               V35          V36          V37          V38          V39
#>  [1,] 0.001388547 0.0006558971 0.0017097371 0.0035434414 0.0049668440
#>  [2,] 0.001527379 0.0010621945 0.0016251603 0.0025868223 0.0034120197
#>  [3,] 0.001246371 0.0010553095 0.0010531912 0.0011511819 0.0012483923
#>  [4,] 0.001024544 0.0008978705 0.0008498019 0.0008635829 0.0008933629
#>  [5,] 0.001803435 0.0017701831 0.0014766476 0.0008738504 0.0006242823
#>  [6,] 0.001234417 0.0008422533 0.0013826323 0.0022795729 0.0030350630
#>  [7,] 0.001723957 0.0013789115 0.0014736467 0.0018455756 0.0020932494
#>  [8,] 0.002175727 0.0015771737 0.0019961608 0.0030728022 0.0036685673
#>  [9,] 0.001974200 0.0019991309 0.0015759404 0.0007877513 0.0004384740
#> [10,] 0.001605815 0.0012284624 0.0013906686 0.0019110391 0.0021842267
#>                  V4          V40          V41           V42           V43
#>  [1,]  1.305466e-03 0.0035616696 0.0037621207  4.696136e-03  4.913972e-03
#>  [2,]  8.351902e-04 0.0025404747 0.0025910726  3.055680e-03  3.217928e-03
#>  [3,]  2.665776e-04 0.0009553042 0.0009454852  9.749113e-04  1.023979e-03
#>  [4,]  1.698245e-04 0.0006782674 0.0006622952  6.425375e-04  6.948396e-04
#>  [5,]  1.839920e-06 0.0007261363 0.0004810954  8.227176e-05  9.583927e-05
#>  [6,]  7.407800e-04 0.0022629323 0.0022739601  2.659614e-03  2.791690e-03
#>  [7,]  4.989738e-04 0.0015088421 0.0015723172  1.748254e-03  1.829574e-03
#>  [8,]  9.513714e-04 0.0024791703 0.0027260854  3.332200e-03  3.505218e-03
#>  [9,] -7.787606e-05 0.0006298260 0.0003228937 -1.835191e-04 -1.704112e-04
#> [10,]  5.415665e-04 0.0014889390 0.0016018067  1.849070e-03  1.959749e-03
#>                 V44           V45           V46           V47           V48
#>  [1,]  4.343009e-03  0.0044384108  3.560085e-03  2.904816e-03  2.295194e-03
#>  [2,]  2.833994e-03  0.0028357572  2.313640e-03  1.899533e-03  1.498343e-03
#>  [3,]  8.459282e-04  0.0007730667  6.857979e-04  5.751910e-04  4.605726e-04
#>  [4,]  5.760096e-04  0.0004987084  4.702014e-04  3.966365e-04  3.188103e-04
#>  [5,]  5.944306e-06 -0.0002021947 -4.822599e-05  2.552398e-05  1.470354e-05
#>  [6,]  2.468369e-03  0.0024830958  2.009961e-03  1.661790e-03  1.317162e-03
#>  [7,]  1.541989e-03  0.0014989853  1.287239e-03  1.048823e-03  8.479451e-04
#>  [8,]  3.064766e-03  0.0031857313  2.649302e-03  2.102834e-03  1.683806e-03
#>  [9,] -2.447796e-04 -0.0005102297 -2.767725e-04 -1.547160e-04 -1.293522e-04
#> [10,]  1.679849e-03  0.0016979895  1.453324e-03  1.162909e-03  9.375357e-04
#>                 V49           V5           V50           V51           V52
#>  [1,]  1.310253e-03 1.925942e-03  4.157644e-04  4.659572e-04  2.922662e-04
#>  [2,]  8.414891e-04 1.258584e-03  2.753793e-04  3.013173e-04  1.900090e-04
#>  [3,]  2.414021e-04 4.347706e-04  9.504675e-05  1.011175e-04  6.544976e-05
#>  [4,]  1.576382e-04 2.914147e-04  6.362909e-05  7.094122e-05  4.591405e-05
#>  [5,] -2.721925e-05 1.082200e-04  2.237760e-05  4.811485e-08  4.188619e-06
#>  [6,]  7.440486e-04 1.110895e-03  2.414127e-04  2.615858e-04  1.630434e-04
#>  [7,]  4.562029e-04 7.577255e-04  1.654416e-04  1.859471e-04  1.199631e-04
#>  [8,]  9.183849e-04 1.340321e-03  3.069861e-04  3.621223e-04  2.299579e-04
#>  [9,] -1.123096e-04 6.003895e-06 -7.652769e-07 -2.950606e-05 -1.369715e-05
#> [10,]  5.015998e-04 7.893955e-04  1.767066e-04  2.053365e-04  1.314256e-04
#>                V53          V54          V55          V56          V57
#>  [1,] 1.958916e-04 2.040582e-04 1.653766e-04 1.427294e-04 1.247949e-04
#>  [2,] 1.316442e-04 1.347536e-04 1.128588e-04 9.614308e-05 8.762501e-05
#>  [3,] 5.304445e-05 5.186441e-05 4.736089e-05 3.927709e-05 3.953668e-05
#>  [4,] 3.746501e-05 3.559926e-05 3.307458e-05 2.897548e-05 2.967340e-05
#>  [5,] 2.018956e-05 1.299414e-05 2.029045e-05 1.539348e-05 2.315104e-05
#>  [6,] 1.137851e-04 1.181705e-04 9.807303e-05 8.585106e-05 7.652587e-05
#>  [7,] 8.844589e-05 9.156234e-05 8.004699e-05 6.612614e-05 6.306034e-05
#>  [8,] 1.550589e-04 1.633183e-04 1.361588e-04 1.153129e-04 1.049446e-04
#>  [9,] 1.046852e-05 1.983439e-06 1.231306e-05 8.088833e-06 1.940597e-05
#> [10,] 9.316172e-05 9.614944e-05 8.302219e-05 7.042466e-05 6.540191e-05
#>                 V58          V59           V6          V60           V7
#>  [1,]  1.430284e-04 1.426388e-04 0.0021192360 1.034919e-04 2.218422e-03
#>  [2,]  9.330435e-05 9.461050e-05 0.0014254161 7.002943e-05 1.473431e-03
#>  [3,]  3.365062e-05 3.818606e-05 0.0005545403 2.719356e-05 5.440142e-04
#>  [4,]  2.208737e-05 2.664376e-05 0.0003952497 1.905467e-05 3.744647e-04
#>  [5,]  5.999567e-06 1.204636e-05 0.0002324874 1.119472e-05 1.725506e-04
#>  [6,]  8.228921e-05 8.380447e-05 0.0012472532 6.210470e-05 1.306833e-03
#>  [7,]  6.057067e-05 6.568486e-05 0.0009303369 4.600227e-05 9.421436e-04
#>  [8,]  1.104056e-04 1.138645e-04 0.0015726483 8.196699e-05 1.653632e-03
#>  [9,] -1.787554e-06 5.651633e-06 0.0001379657 6.828529e-06 6.186007e-05
#> [10,]  6.360318e-05 6.831227e-05 0.0009534158 4.911915e-05 9.778674e-04
#>                 V8            V9
#>  [1,] 2.616215e-03  3.896507e-03
#>  [2,] 1.741186e-03  2.541299e-03
#>  [3,] 6.443463e-04  8.292375e-04
#>  [4,] 4.594834e-04  5.716016e-04
#>  [5,] 1.704248e-04  5.227068e-05
#>  [6,] 1.553901e-03  2.279639e-03
#>  [7,] 1.114622e-03  1.510928e-03
#>  [8,] 1.991265e-03  2.879123e-03
#>  [9,] 4.221085e-05 -1.751828e-04
#> [10,] 1.182729e-03  1.657522e-03
#> 
#> $vW[[2]]
#>             [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#> [1,] -0.06490562 -0.06108619 -0.1522115 -0.1589867 -0.09329969 -0.05437122
#> [2,]  0.06490562  0.06108619  0.1522115  0.1589867  0.09329969  0.05437122
#>            [,7]       [,8]       [,9]      [,10]
#> [1,] -0.1374165 -0.1384469 -0.1094356 -0.1458482
#> [2,]  0.1374165  0.1384469  0.1094356  0.1458482
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.329040015 0.171708547 0.657362349 0.838088608 0.381697421 0.000879513
#>  [7] 0.548892603 0.698195118 0.283218311 0.642876848
#> 
#> $B[[2]]
#> [1]  0.1847949 -0.2144595
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.014877124 0.010515147 0.004259410 0.003096317 0.002848812 0.009184927
#>  [7] 0.006832790 0.011176711 0.002291595 0.006876571
#> 
#> $vB[[2]]
#> [1] -0.1836773  0.1836773
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#>  [2,] 0.0137 0.0874 0.1100 0.1084 0.1094 0.1023 0.0601 0.0906 0.1313 0.2758
#>  [3,] 0.0221 0.1810 0.2549 0.2984 0.2624 0.1893 0.0668 0.2666 0.4274 0.6291
#>  [4,] 0.0047 0.3469 0.3265 0.3263 0.2301 0.1253 0.2102 0.2401 0.1928 0.1673
#>  [5,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#>  [6,] 0.0094 0.5079 0.3350 0.0834 0.3004 0.3957 0.3769 0.3828 0.1247 0.1363
#>  [7,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#>  [8,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#>  [9,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [10,] 0.0442 0.1268 0.1109 0.2375 0.2007 0.2140 0.1109 0.2036 0.2468 0.6682
#> [11,] 0.0228 0.2227 0.2621 0.3109 0.2859 0.3316 0.3755 0.4499 0.4765 0.6254
#> [12,] 0.0181 0.1045 0.0904 0.0671 0.0997 0.1056 0.0346 0.1231 0.1626 0.3652
#> [13,] 0.0114 0.1799 0.2486 0.2973 0.3672 0.4394 0.5258 0.6755 0.7402 0.8284
#> [14,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [15,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#> [16,] 0.0188 0.1806 0.2139 0.1523 0.1975 0.4844 0.7298 0.7807 0.7906 0.6122
#> [17,] 0.0629 0.3597 0.5466 0.5205 0.5127 0.5395 0.6558 0.8705 0.9786 0.9335
#> [18,] 0.0129 0.2245 0.2431 0.3134 0.3206 0.2917 0.2249 0.2347 0.2143 0.2939
#> [19,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [20,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [21,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [22,] 0.1371 0.2067 0.4257 0.5484 0.7131 0.7003 0.6777 0.7939 0.9382 0.8925
#> [23,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [24,] 0.0164 0.0251 0.0801 0.1056 0.1266 0.0890 0.0198 0.1133 0.2826 0.3234
#> [25,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#> [26,] 0.0264 0.0945 0.1132 0.0840 0.0717 0.1968 0.2633 0.4191 0.5050 0.6711
#> [27,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#> [28,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [29,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [30,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [31,] 0.0201 0.1097 0.0841 0.0942 0.1204 0.0420 0.0031 0.0162 0.0624 0.2127
#> [32,] 0.0095 0.3745 0.4229 0.4499 0.5404 0.4303 0.3333 0.3496 0.3426 0.2851
#> [33,] 0.0210 0.0686 0.1125 0.1741 0.2710 0.3087 0.3575 0.4998 0.6011 0.6470
#> [34,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#> [35,] 0.0176 0.0474 0.0526 0.1854 0.1040 0.0948 0.0912 0.1688 0.1568 0.0375
#> [36,] 0.0365 0.3735 0.3042 0.0387 0.2679 0.5397 0.6204 0.7257 0.8350 0.6888
#> [37,] 0.0599 0.0351 0.1582 0.2023 0.2268 0.2829 0.3819 0.4665 0.6687 0.8647
#> [38,] 0.0211 0.1345 0.0785 0.0367 0.1227 0.2614 0.4280 0.6122 0.7435 0.8130
#> [39,] 0.0274 0.1808 0.2366 0.0906 0.1749 0.4012 0.5187 0.7312 0.9062 0.9260
#> [40,] 0.0293 0.0860 0.0414 0.0472 0.0835 0.0938 0.1466 0.0809 0.1179 0.2179
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#>  [2,] 0.3660 0.0297 0.5269 0.5810 0.6181 0.5875 0.4639 0.5424 0.7367 0.9089
#>  [3,] 0.7782 0.0065 0.7686 0.8099 0.8493 0.9440 0.9450 0.9655 0.8045 0.4969
#>  [4,] 0.1228 0.0059 0.0902 0.1557 0.3291 0.5268 0.6740 0.7906 0.8938 0.9395
#>  [5,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#>  [6,] 0.2678 0.0611 0.9188 0.9779 0.3236 0.1944 0.1874 0.0885 0.3443 0.2953
#>  [7,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#>  [8,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#>  [9,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [10,] 0.8345 0.0477 0.8252 0.8017 0.8982 0.9664 0.8515 0.6626 0.3241 0.2054
#> [11,] 0.7304 0.0106 0.8702 0.9349 0.9614 0.9126 0.9443 1.0000 0.9455 0.8815
#> [12,] 0.3262 0.0146 0.2995 0.2109 0.2104 0.2085 0.2282 0.0747 0.1969 0.4086
#> [13,] 0.9033 0.0222 0.9584 1.0000 0.9982 0.8899 0.7493 0.6367 0.6744 0.7207
#> [14,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [15,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#> [16,] 0.4200 0.0370 0.2807 0.5148 0.7569 0.8596 1.0000 0.8457 0.6797 0.6971
#> [17,] 0.7917 0.1065 0.7383 0.6908 0.3850 0.0671 0.0502 0.2717 0.2839 0.2234
#> [18,] 0.4898 0.0141 0.6127 0.7531 0.7718 0.7432 0.8673 0.9308 0.9836 1.0000
#> [19,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [20,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [21,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [22,] 0.9146 0.1226 0.7832 0.7960 0.7983 0.7716 0.6615 0.4860 0.5572 0.4697
#> [23,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [24,] 0.3238 0.0173 0.4333 0.6068 0.7652 0.9203 0.9719 0.9207 0.7545 0.8289
#> [25,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#> [26,] 0.7922 0.0071 0.8381 0.8759 0.9422 1.0000 0.9931 0.9575 0.8647 0.7215
#> [27,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#> [28,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [29,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [30,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [31,] 0.3436 0.0026 0.3813 0.3825 0.4764 0.6313 0.7523 0.8675 0.8788 0.7901
#> [32,] 0.4062 0.0308 0.6833 0.7650 0.6670 0.5703 0.5995 0.6484 0.8614 0.9819
#> [33,] 0.8067 0.0121 0.9008 0.8906 0.9338 1.0000 0.9102 0.8496 0.7867 0.7688
#> [34,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#> [35,] 0.1316 0.0172 0.2086 0.1976 0.0946 0.1965 0.1242 0.0616 0.2141 0.4642
#> [36,] 0.4450 0.1632 0.3921 0.5605 0.7545 0.8311 1.0000 0.8762 0.7092 0.7009
#> [37,] 0.9361 0.0474 0.9367 0.9144 0.9162 0.9311 0.8604 0.7327 0.5763 0.4162
#> [38,] 0.9006 0.0128 0.9603 0.9162 0.9140 0.7851 0.5134 0.3439 0.3290 0.2571
#> [39,] 0.7434 0.0242 0.4463 0.5103 0.6952 0.7755 0.8364 0.7283 0.6399 0.5759
#> [40,] 0.3326 0.0644 0.3258 0.2111 0.2302 0.3361 0.4259 0.4609 0.2606 0.0874
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#>  [2,] 1.0000 0.8247 0.0116 0.5441 0.3349 0.0877 0.1600 0.4169 0.6576 0.7390
#>  [3,] 0.3960 0.3856 0.0164 0.5574 0.7309 0.8549 0.9425 0.8726 0.6673 0.4694
#>  [4,] 0.9493 0.9040 0.0080 0.9151 0.8828 0.8086 0.7180 0.6720 0.6447 0.6879
#>  [5,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#>  [6,] 0.5908 0.4564 0.1136 0.7334 0.1969 0.2790 0.6212 0.8681 0.8621 0.9380
#>  [7,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#>  [8,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#>  [9,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [10,] 0.5669 0.5726 0.0049 0.4877 0.7532 0.7600 0.5185 0.4120 0.5560 0.5569
#> [11,] 0.7520 0.7068 0.0130 0.5986 0.3857 0.2510 0.2162 0.0968 0.1323 0.1344
#> [12,] 0.6385 0.7970 0.0026 0.7508 0.5517 0.2214 0.4672 0.4479 0.2297 0.3235
#> [13,] 0.6821 0.5512 0.0269 0.4789 0.3924 0.2533 0.1089 0.1390 0.2551 0.3301
#> [14,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [15,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#> [16,] 0.5843 0.4772 0.0953 0.5201 0.4241 0.1592 0.1668 0.0588 0.3967 0.7147
#> [17,] 0.1911 0.0408 0.1526 0.2531 0.1979 0.1891 0.2433 0.1956 0.2667 0.1340
#> [18,] 0.9595 0.8722 0.0309 0.6862 0.4901 0.3280 0.3115 0.1969 0.1019 0.0317
#> [19,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [20,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [21,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [22,] 0.5640 0.4517 0.1385 0.3369 0.2684 0.2339 0.3052 0.3016 0.2753 0.1041
#> [23,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [24,] 0.8907 0.7309 0.0347 0.6896 0.5829 0.4935 0.3101 0.0306 0.0244 0.1108
#> [25,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#> [26,] 0.5801 0.4964 0.0342 0.4886 0.4079 0.2443 0.1768 0.2472 0.3518 0.3762
#> [27,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#> [28,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [29,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [30,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [31,] 0.8357 0.9631 0.0138 0.9619 0.9236 0.8903 0.9708 0.9647 0.7892 0.5307
#> [32,] 0.9380 0.8435 0.0539 0.6074 0.5403 0.6890 0.5977 0.3244 0.0516 0.3157
#> [33,] 0.7718 0.6268 0.0203 0.4301 0.2077 0.1198 0.1660 0.2618 0.3862 0.3958
#> [34,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#> [35,] 0.6471 0.6340 0.0501 0.6107 0.7046 0.5376 0.5934 0.8443 0.9481 0.9705
#> [36,] 0.5014 0.3942 0.1636 0.4456 0.4072 0.0773 0.1423 0.0401 0.3597 0.6847
#> [37,] 0.4113 0.4146 0.0498 0.3149 0.2936 0.3169 0.3149 0.4132 0.3994 0.4195
#> [38,] 0.3685 0.5765 0.0015 0.6190 0.4613 0.3615 0.4434 0.3864 0.3093 0.2138
#> [39,] 0.4146 0.3495 0.0621 0.4437 0.2665 0.2024 0.1942 0.0765 0.3725 0.5843
#> [40,] 0.2862 0.5606 0.0390 0.8344 0.8096 0.7250 0.8048 0.9435 1.0000 0.8960
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#>  [2,] 0.7963 0.7493 0.6795 0.0082 0.4713 0.2355 0.1704 0.2728 0.4016 0.4125
#>  [3,] 0.1546 0.1748 0.3607 0.0487 0.5208 0.5177 0.3702 0.2240 0.0816 0.0395
#>  [4,] 0.6241 0.4936 0.4144 0.0554 0.4240 0.4546 0.4392 0.4323 0.4921 0.4710
#>  [5,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#>  [6,] 0.8327 0.9480 0.6721 0.1203 0.4436 0.5163 0.3809 0.1557 0.1449 0.2662
#>  [7,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#>  [8,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#>  [9,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [10,] 0.1336 0.3831 0.4611 0.0581 0.4330 0.2556 0.1466 0.3489 0.2659 0.0944
#> [11,] 0.2250 0.3244 0.3939 0.0842 0.3806 0.3258 0.3654 0.2983 0.1779 0.1535
#> [12,] 0.4480 0.5581 0.6520 0.0141 0.5354 0.2478 0.2268 0.1788 0.0898 0.0536
#> [13,] 0.2818 0.2142 0.2266 0.0384 0.2142 0.2354 0.2871 0.2596 0.1925 0.1256
#> [14,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [15,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#> [16,] 0.7319 0.3509 0.0589 0.0824 0.2690 0.4200 0.3874 0.2440 0.2000 0.2307
#> [17,] 0.1073 0.2023 0.1794 0.1229 0.0227 0.1313 0.1775 0.1549 0.1626 0.0708
#> [18,] 0.0756 0.0907 0.1066 0.0375 0.1380 0.0665 0.1475 0.2470 0.2788 0.2709
#> [19,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [20,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [21,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [22,] 0.1757 0.3156 0.3603 0.1484 0.2736 0.1301 0.2458 0.3404 0.1753 0.0679
#> [23,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [24,] 0.1594 0.1371 0.0696 0.0070 0.0452 0.0620 0.1421 0.1597 0.1384 0.0372
#> [25,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#> [26,] 0.2909 0.2311 0.3168 0.0793 0.3554 0.3741 0.4443 0.3261 0.1963 0.0864
#> [27,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#> [28,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [29,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [30,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [31,] 0.2718 0.1953 0.1374 0.0062 0.3105 0.3790 0.4105 0.3355 0.2998 0.2748
#> [32,] 0.3590 0.3881 0.5716 0.0411 0.4314 0.3051 0.4393 0.4302 0.4831 0.5084
#> [33,] 0.3248 0.2302 0.3250 0.1036 0.4022 0.4344 0.4008 0.3370 0.2518 0.2101
#> [34,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#> [35,] 0.7766 0.6313 0.5760 0.0285 0.6148 0.5450 0.4813 0.3406 0.1916 0.1134
#> [36,] 0.7076 0.3597 0.0612 0.1421 0.3027 0.3966 0.3868 0.2380 0.2059 0.2288
#> [37,] 0.4532 0.4419 0.4737 0.0387 0.3431 0.3194 0.3370 0.2493 0.2650 0.1748
#> [38,] 0.1112 0.1386 0.1523 0.0450 0.0996 0.1644 0.1902 0.1313 0.1776 0.2000
#> [39,] 0.4827 0.2347 0.0999 0.0560 0.3244 0.3990 0.2975 0.1684 0.1761 0.1683
#> [40,] 0.5516 0.3037 0.2338 0.0173 0.2382 0.3318 0.3821 0.1575 0.2228 0.1582
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#>  [2,] 0.3470 0.2739 0.1790 0.0922 0.0241 0.0276 0.0169 0.0081 0.0040 0.0025
#>  [3,] 0.0785 0.1052 0.1034 0.0764 0.0519 0.0216 0.0167 0.0089 0.0051 0.0015
#>  [4,] 0.3196 0.2241 0.1806 0.0990 0.0883 0.0251 0.0129 0.0095 0.0126 0.0069
#>  [5,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#>  [6,] 0.1806 0.1699 0.2559 0.1129 0.0403 0.0201 0.0480 0.0234 0.0175 0.0352
#>  [7,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#>  [8,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#>  [9,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [10,] 0.1370 0.1344 0.0416 0.0719 0.0278 0.0637 0.0210 0.0204 0.0216 0.0135
#> [11,] 0.1199 0.0959 0.0765 0.0649 0.1117 0.0313 0.0185 0.0098 0.0178 0.0077
#> [12,] 0.0374 0.0990 0.0956 0.0317 0.0421 0.0142 0.0076 0.0223 0.0255 0.0145
#> [13,] 0.1003 0.0951 0.1210 0.0728 0.1217 0.0174 0.0213 0.0269 0.0152 0.0257
#> [14,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [15,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#> [16,] 0.1886 0.1960 0.1701 0.1366 0.0249 0.0398 0.0143 0.0093 0.0033 0.0113
#> [17,] 0.0129 0.0795 0.0762 0.0117 0.1437 0.0061 0.0257 0.0089 0.0262 0.0108
#> [18,] 0.2283 0.1818 0.1185 0.0546 0.0767 0.0219 0.0204 0.0124 0.0093 0.0072
#> [19,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [20,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [21,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [22,] 0.1062 0.0643 0.0532 0.0531 0.1776 0.0272 0.0171 0.0118 0.0129 0.0344
#> [23,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [24,] 0.0688 0.0867 0.0513 0.0092 0.0187 0.0198 0.0118 0.0090 0.0223 0.0179
#> [25,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#> [26,] 0.1688 0.1991 0.1217 0.0628 0.1043 0.0323 0.0253 0.0214 0.0262 0.0177
#> [27,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#> [28,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [29,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [30,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [31,] 0.2024 0.1043 0.0453 0.0337 0.0133 0.0122 0.0072 0.0108 0.0070 0.0063
#> [32,] 0.1952 0.1539 0.2037 0.1054 0.0613 0.0251 0.0357 0.0181 0.0019 0.0102
#> [33,] 0.1181 0.1150 0.0550 0.0293 0.1675 0.0183 0.0104 0.0117 0.0101 0.0061
#> [34,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#> [35,] 0.0640 0.0911 0.0980 0.0563 0.0262 0.0187 0.0088 0.0042 0.0175 0.0171
#> [36,] 0.1704 0.1587 0.1792 0.1022 0.1130 0.0151 0.0223 0.0110 0.0071 0.0205
#> [37,] 0.0932 0.0530 0.0081 0.0342 0.1026 0.0137 0.0028 0.0013 0.0005 0.0227
#> [38,] 0.0765 0.0727 0.0749 0.0449 0.0711 0.0134 0.0174 0.0117 0.0023 0.0047
#> [39,] 0.0729 0.1190 0.1297 0.0748 0.1129 0.0067 0.0255 0.0113 0.0108 0.0085
#> [40,] 0.1433 0.1634 0.1133 0.0567 0.0476 0.0133 0.0170 0.0035 0.0052 0.0083
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#>  [2,] 0.0036 0.0058 0.0067 0.0035 0.0043 0.0253 0.0033 0.0279 0.0130 0.0489
#>  [3,] 0.0075 0.0058 0.0016 0.0070 0.0074 0.0849 0.0038 0.0812 0.1833 0.2228
#>  [4,] 0.0039 0.0068 0.0060 0.0045 0.0002 0.1278 0.0029 0.1674 0.1373 0.2922
#>  [5,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#>  [6,] 0.0158 0.0326 0.0201 0.0168 0.0245 0.1227 0.0154 0.2495 0.4566 0.6587
#>  [7,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#>  [8,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#>  [9,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [10,] 0.0055 0.0073 0.0080 0.0105 0.0059 0.0678 0.0105 0.1664 0.1490 0.0974
#> [11,] 0.0074 0.0095 0.0055 0.0045 0.0063 0.1506 0.0039 0.1776 0.0997 0.1428
#> [12,] 0.0233 0.0041 0.0018 0.0048 0.0089 0.0473 0.0085 0.0361 0.0741 0.1398
#> [13,] 0.0097 0.0041 0.0050 0.0145 0.0103 0.2062 0.0025 0.1489 0.0929 0.1350
#> [14,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [15,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#> [16,] 0.0030 0.0057 0.0090 0.0057 0.0068 0.0488 0.0024 0.1424 0.1972 0.1873
#> [17,] 0.0138 0.0187 0.0230 0.0057 0.0113 0.1190 0.0131 0.0884 0.0907 0.2107
#> [18,] 0.0019 0.0027 0.0054 0.0017 0.0024 0.0787 0.0029 0.0662 0.1108 0.1777
#> [19,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [20,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [21,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [22,] 0.0065 0.0067 0.0022 0.0079 0.0146 0.1428 0.0051 0.1773 0.2161 0.1630
#> [23,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [24,] 0.0084 0.0068 0.0032 0.0035 0.0056 0.0671 0.0040 0.1056 0.0697 0.0962
#> [25,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#> [26,] 0.0037 0.0068 0.0121 0.0077 0.0078 0.0783 0.0066 0.1417 0.1176 0.0453
#> [27,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#> [28,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [29,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [30,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [31,] 0.0030 0.0011 0.0007 0.0024 0.0057 0.0151 0.0044 0.0541 0.0210 0.0505
#> [32,] 0.0133 0.0040 0.0042 0.0030 0.0031 0.1039 0.0033 0.1016 0.1394 0.2592
#> [33,] 0.0031 0.0099 0.0080 0.0107 0.0161 0.0418 0.0133 0.0723 0.0828 0.0494
#> [34,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#> [35,] 0.0079 0.0050 0.0112 0.0179 0.0294 0.0351 0.0063 0.0362 0.0535 0.0258
#> [36,] 0.0164 0.0063 0.0078 0.0094 0.0110 0.1306 0.0068 0.2112 0.2268 0.2992
#> [37,] 0.0209 0.0081 0.0117 0.0114 0.0112 0.0773 0.0100 0.0853 0.0447 0.1094
#> [38,] 0.0049 0.0031 0.0024 0.0039 0.0051 0.1563 0.0015 0.1518 0.1206 0.1666
#> [39,] 0.0047 0.0074 0.0104 0.0161 0.0220 0.0973 0.0173 0.1823 0.1745 0.1440
#> [40,] 0.0078 0.0075 0.0105 0.0160 0.0095 0.0816 0.0011 0.0993 0.0315 0.0736
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4689679 0.4918167 0.8638585 0.8816103 0.6709671 0.3969336 0.8227821
#>  [2,] 0.4169799 0.4324848 0.8732682 0.8707085 0.6195477 0.3826001 0.8106469
#>  [3,] 0.3960206 0.4105217 0.8956740 0.9077722 0.5643669 0.3836776 0.8228955
#>  [4,] 0.3400029 0.4068158 0.8805342 0.8526433 0.6188324 0.2977424 0.7980924
#>  [5,] 0.2682295 0.2834528 0.7687597 0.7282972 0.4961689 0.1966358 0.6197945
#>  [6,] 0.2392709 0.2416196 0.7927797 0.7626518 0.4364914 0.2601927 0.6203822
#>  [7,] 0.3554058 0.3798429 0.8699278 0.8902417 0.5432625 0.3607285 0.8137223
#>  [8,] 0.4604946 0.4599915 0.8966201 0.9213706 0.6287582 0.3864866 0.8516546
#>  [9,] 0.3424837 0.2914111 0.7359239 0.7820170 0.3802014 0.2531096 0.5969960
#> [10,] 0.4213587 0.4149666 0.8570763 0.8853358 0.5538001 0.3642102 0.7683017
#> [11,] 0.4294240 0.4113130 0.8791374 0.9136793 0.5852417 0.3733764 0.8349139
#> [12,] 0.4577747 0.4427232 0.8043219 0.8143810 0.6167696 0.3715586 0.7058519
#> [13,] 0.4227964 0.3792156 0.8564029 0.8914328 0.5217472 0.3400555 0.7812809
#> [14,] 0.3559334 0.3890682 0.8405015 0.7727346 0.5728327 0.3066427 0.6882094
#> [15,] 0.4960152 0.4996560 0.8813336 0.9123193 0.6631822 0.4352627 0.8498218
#> [16,] 0.3690720 0.3576428 0.8532868 0.8525419 0.5214388 0.3181929 0.7814467
#> [17,] 0.3715766 0.3171934 0.7064570 0.7346460 0.4025016 0.2669477 0.5655647
#> [18,] 0.4789954 0.4888659 0.8793871 0.9116891 0.6589740 0.4127063 0.8468604
#> [19,] 0.4281554 0.4493050 0.8804669 0.8945885 0.6474344 0.3652043 0.8230740
#> [20,] 0.4425707 0.4313241 0.8754696 0.9085199 0.5872994 0.3772602 0.8298380
#> [21,] 0.4997557 0.5260209 0.8356595 0.8554839 0.6653705 0.4475399 0.8124491
#> [22,] 0.3314014 0.3153851 0.8051705 0.8414959 0.4414724 0.2634885 0.6774143
#> [23,] 0.4870290 0.4914638 0.8624180 0.8853176 0.6462822 0.4585500 0.8281138
#> [24,] 0.5516325 0.5590815 0.8825395 0.9141822 0.6873202 0.4822922 0.8610744
#> [25,] 0.4553775 0.4887514 0.8644461 0.8834331 0.6403519 0.4066293 0.8248343
#> [26,] 0.4336677 0.4073884 0.8771129 0.9032389 0.5744304 0.3899575 0.8259759
#> [27,] 0.4090584 0.3898209 0.8352843 0.8299152 0.5470488 0.3095465 0.7170285
#> [28,] 0.3245728 0.3421917 0.8111420 0.7945636 0.5094603 0.2939464 0.6958468
#> [29,] 0.3643016 0.3777569 0.8930219 0.8840734 0.5122518 0.3481776 0.8053581
#> [30,] 0.3814524 0.4126466 0.8590174 0.8021800 0.6103621 0.3172527 0.7199632
#> [31,] 0.4744570 0.5181013 0.9124262 0.9108804 0.7022264 0.4015590 0.8551014
#> [32,] 0.3484701 0.3700709 0.8623254 0.8763679 0.5621647 0.2901836 0.7825367
#> [33,] 0.4319706 0.3987591 0.8822224 0.9067253 0.5682419 0.3771215 0.8191697
#> [34,] 0.3327273 0.3125465 0.8180258 0.8748976 0.4741182 0.2924585 0.7434051
#> [35,] 0.3827098 0.4110441 0.8411814 0.7768522 0.5832773 0.3306180 0.7021598
#> [36,] 0.3320430 0.3144674 0.8305541 0.8348613 0.4812123 0.2941728 0.7538028
#> [37,] 0.4130071 0.3785258 0.8605752 0.8898227 0.5053011 0.3609173 0.7715135
#> [38,] 0.4688571 0.3902887 0.8277137 0.8801142 0.5248619 0.3616072 0.7276918
#> [39,] 0.4052914 0.3642709 0.8304817 0.8479958 0.5023079 0.3316319 0.7521919
#> [40,] 0.4090206 0.4329135 0.8347144 0.8076069 0.5899191 0.3594543 0.7237826
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7984571 0.6465934 0.8334392
#>  [2,] 0.7698837 0.5989924 0.8085111
#>  [3,] 0.8530877 0.6250747 0.8510715
#>  [4,] 0.7735658 0.4900475 0.7906627
#>  [5,] 0.6163139 0.3530379 0.6521231
#>  [6,] 0.6380589 0.3484204 0.6409860
#>  [7,] 0.7989514 0.6300121 0.8239453
#>  [8,] 0.8558723 0.7205989 0.8821887
#>  [9,] 0.6969757 0.5169715 0.6838717
#> [10,] 0.8171187 0.6189879 0.8253052
#> [11,] 0.8375853 0.7041667 0.8682778
#> [12,] 0.7230341 0.5370472 0.7504953
#> [13,] 0.8159430 0.6787411 0.8387848
#> [14,] 0.7068300 0.4094638 0.7073422
#> [15,] 0.8271789 0.7180359 0.8665138
#> [16,] 0.8030365 0.5992796 0.7981622
#> [17,] 0.6551648 0.4824112 0.6248759
#> [18,] 0.8240420 0.7082411 0.8647178
#> [19,] 0.8149628 0.6331055 0.8438396
#> [20,] 0.8397874 0.7105258 0.8631528
#> [21,] 0.7909391 0.6490640 0.8083355
#> [22,] 0.7455081 0.5649422 0.7494323
#> [23,] 0.8190124 0.6724641 0.8385742
#> [24,] 0.8537847 0.7473124 0.8810699
#> [25,] 0.8168643 0.6476481 0.8313032
#> [26,] 0.8426000 0.6940161 0.8630861
#> [27,] 0.7459028 0.5563927 0.7690942
#> [28,] 0.7337867 0.4475645 0.7402103
#> [29,] 0.8235672 0.5759282 0.8153023
#> [30,] 0.7313094 0.4436280 0.7433893
#> [31,] 0.8501886 0.6246349 0.8635867
#> [32,] 0.7604333 0.5584766 0.8055428
#> [33,] 0.8362020 0.6937350 0.8613635
#> [34,] 0.7920455 0.6423299 0.8093796
#> [35,] 0.7141109 0.4338811 0.7136065
#> [36,] 0.7734450 0.5542262 0.7688913
#> [37,] 0.8248811 0.6502849 0.8312043
#> [38,] 0.8114144 0.6468463 0.8180193
#> [39,] 0.8004699 0.6066779 0.7877438
#> [40,] 0.7617733 0.4669239 0.7300609
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.2827839 0.7172161
#>  [2,] 0.2917985 0.7082015
#>  [3,] 0.2862553 0.7137447
#>  [4,] 0.3017628 0.6982372
#>  [5,] 0.3411041 0.6588959
#>  [6,] 0.3401620 0.6598380
#>  [7,] 0.2962285 0.7037715
#>  [8,] 0.2780053 0.7219947
#>  [9,] 0.3281372 0.6718628
#> [10,] 0.2910818 0.7089182
#> [11,] 0.2848910 0.7151090
#> [12,] 0.2995774 0.7004226
#> [13,] 0.2926718 0.7073282
#> [14,] 0.3137705 0.6862295
#> [15,] 0.2750925 0.7249075
#> [16,] 0.3005969 0.6994031
#> [17,] 0.3317994 0.6682006
#> [18,] 0.2777582 0.7222418
#> [19,] 0.2863313 0.7136687
#> [20,] 0.2835038 0.7164962
#> [21,] 0.2802570 0.7197430
#> [22,] 0.3167757 0.6832243
#> [23,] 0.2779064 0.7220936
#> [24,] 0.2653835 0.7346165
#> [25,] 0.2823242 0.7176758
#> [26,] 0.2845194 0.7154806
#> [27,] 0.3041360 0.6958640
#> [28,] 0.3156628 0.6843372
#> [29,] 0.2948730 0.7051270
#> [30,] 0.3059980 0.6940020
#> [31,] 0.2747107 0.7252893
#> [32,] 0.3039851 0.6960149
#> [33,] 0.2859340 0.7140660
#> [34,] 0.3078003 0.6921997
#> [35,] 0.3089811 0.6910189
#> [36,] 0.3102031 0.6897969
#> [37,] 0.2924365 0.7075635
#> [38,] 0.2918521 0.7081479
#> [39,] 0.2995269 0.7004731
#> [40,] 0.3007839 0.6992161
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>               [,1]        [,2]      [,3]      [,4]         [,5]        [,6]
#>  [1,] -0.124287971 -0.03273599 1.8477144 2.0077682  0.712562152 -0.41825849
#>  [2,] -0.335183709 -0.27172024 1.9301699 1.9072376  0.487628840 -0.47852665
#>  [3,] -0.422073784 -0.36180892 2.1500557 2.2867324  0.258904260 -0.47396751
#>  [4,] -0.663281082 -0.37714462 1.9974984 1.7554852  0.484595508 -0.85807177
#>  [5,] -1.003623914 -0.92739834 1.2013210 0.9860006 -0.015324891 -1.40745482
#>  [6,] -1.156680738 -1.14382056 1.3417629 1.1672731 -0.255414078 -1.04496707
#>  [7,] -0.595360610 -0.49021498 1.9003207 2.0932122  0.173483869 -0.57220377
#>  [8,] -0.158351811 -0.16037671 2.1602221 2.4611161  0.526893139 -0.46210485
#>  [9,] -0.652245578 -0.88854077 1.0248895 1.2774592 -0.488693678 -1.08209607
#> [10,] -0.317198178 -0.34347087 1.7912158 2.0439593  0.216036838 -0.55713702
#> [11,] -0.284201674 -0.35854030 1.9842865 2.3594100  0.344329029 -0.51775896
#> [12,] -0.169304345 -0.23011730 1.4135284 1.4787324  0.475859040 -0.52553625
#> [13,] -0.311304131 -0.49287879 1.7857294 2.1054604  0.087043546 -0.66304678
#> [14,] -0.593058562 -0.45123079 1.6619643 1.2238170  0.293417970 -0.81586225
#> [15,] -0.015939380 -0.00137599 2.0051204 2.3422887  0.677507721 -0.26041074
#> [16,] -0.536200008 -0.58560963 1.7606160 1.7546783  0.085807635 -0.76208915
#> [17,] -0.525458969 -0.76669995 0.8782383 1.0183238 -0.395052472 -1.01016434
#> [18,] -0.084067757 -0.04454367 1.9866392 2.3344355  0.658725528 -0.35278901
#> [19,] -0.289380880 -0.20347903 1.9968588 2.1384917  0.607780956 -0.55284640
#> [20,] -0.230735390 -0.27645082 1.9502109 2.2956959  0.352812280 -0.50119351
#> [21,] -0.000977354  0.10417789 1.6262807 1.7782760  0.687320233 -0.21061555
#> [22,] -0.701853450 -0.77506185 1.4189290 1.6694010 -0.235188375 -1.02791537
#> [23,] -0.051895550 -0.03414815 1.8355201 2.0437797  0.602736969 -0.16618132
#> [24,]  0.207269051  0.23743521 2.0167016 2.3658030  0.787620452 -0.07086101
#> [25,] -0.178965961 -0.04500218 1.8527195 2.0253503  0.576891625 -0.37791721
#> [26,] -0.266902636 -0.37477218 1.9653697 2.2337424  0.299950580 -0.44749105
#> [27,] -0.367859445 -0.44806527 1.6235512 1.5850263  0.188753618 -0.80224038
#> [28,] -0.732835306 -0.65354263 1.4574479 1.3526564  0.037845890 -0.87629359
#> [29,] -0.556742264 -0.49908003 2.1219872 2.0315825  0.049017004 -0.62705938
#> [30,] -0.483387927 -0.35303506 1.8071524 1.3999756  0.448834460 -0.76642638
#> [31,] -0.102261054  0.07243704 2.3436258 2.3244326  0.857922262 -0.39897363
#> [32,] -0.625770924 -0.53191274 1.8347397 1.9584760  0.249952017 -0.89449232
#> [33,] -0.273815597 -0.41063808 2.0136456 2.2742901  0.274681807 -0.50178413
#> [34,] -0.695875782 -0.78824114 1.5030288 1.9449742 -0.103619845 -0.88347374
#> [35,] -0.478062048 -0.35965080 1.6670444 1.2474157  0.336241716 -0.70539149
#> [36,] -0.698959516 -0.77931548 1.5895591 1.6204801 -0.075186085 -0.87520288
#> [37,] -0.351548035 -0.49581008 1.8200757 2.0889311  0.021205102 -0.57138495
#> [38,] -0.124733099 -0.44609871 1.5695099 1.9935123  0.099529711 -0.56839512
#> [39,] -0.383465055 -0.55687450 1.5890453 1.7189674  0.009231525 -0.70081358
#> [40,] -0.368015729 -0.26997406 1.6194147 1.4345349  0.363630860 -0.57773350
#>            [,7]      [,8]        [,9]     [,10]
#>  [1,] 1.5353112 1.3766789  0.60409863 1.6102002
#>  [2,] 1.4542188 1.2076544  0.40126843 1.4403643
#>  [3,] 1.5360893 1.7590269  0.51114426 1.7430295
#>  [4,] 1.3744142 1.2285561 -0.03981519 1.3289244
#>  [5,] 0.4886760 0.4739315 -0.60571217 0.6283845
#>  [6,] 0.4911709 0.5669489 -0.62598980 0.5796462
#>  [7,] 1.4743804 1.3797534  0.53226862 1.5433095
#>  [8,] 1.7476376 1.7814212  0.94743413 2.0133222
#>  [9,] 0.3929640 0.8329374  0.06791208 0.7716223
#> [10,] 1.1987463 1.4969468  0.48525461 1.5527129
#> [11,] 1.6208617 1.6403698  0.86721905 1.8858168
#> [12,] 0.8753219 0.9595618  0.14846084 1.1012558
#> [13,] 1.2731463 1.4890990  0.74799260 1.6492138
#> [14,] 0.7917612 0.8800375 -0.36618244 0.8825104
#> [15,] 1.7332042 1.5657641  0.93473971 1.8704801
#> [16,] 1.2741168 1.4053816  0.40246452 1.3748472
#> [17,] 0.2637777 0.6418202 -0.07038432 0.5102960
#> [18,] 1.7101857 1.5439763  0.88685682 1.8550403
#> [19,] 1.5373146 1.4825853  0.54556260 1.6870786
#> [20,] 1.5844799 1.6566474  0.89793891 1.8417266
#> [21,] 1.4660030 1.3305956  0.61492730 1.4392307
#> [22,] 0.7419143 1.0747968  0.26124437 1.0955868
#> [23,] 1.5723182 1.5096704  0.71935083 1.6476572
#> [24,] 1.8242421 1.7645992  1.08432924 2.0026009
#> [25,] 1.5494499 1.4952457  0.60871707 1.5948918
#> [26,] 1.5573715 1.6777021  0.81896285 1.8411618
#> [27,] 0.9297693 1.0768786  0.22653448 1.2032038
#> [28,] 0.8275978 1.0139204 -0.21051591 1.0470619
#> [29,] 1.4201253 1.5407054  0.30608039 1.4848388
#> [30,] 0.9442790 1.0012759 -0.22645094 1.0636599
#> [31,] 1.7751858 1.7360811  0.50926833 1.8454047
#> [32,] 1.2805108 1.1550564  0.23498159 1.4213046
#> [33,] 1.5107325 1.6302362  0.81763936 1.8266609
#> [34,] 1.0637424 1.3372995  0.58549088 1.4459842
#> [35,] 0.8576041 0.9154341 -0.26603352 0.9129647
#> [36,] 1.1189981 1.2278667  0.21776139 1.2020613
#> [37,] 1.2168768 1.5497744  0.62029169 1.5941863
#> [38,] 0.9829432 1.4592270  0.60520526 1.5029852
#> [39,] 1.1103369 1.3892337  0.43336930 1.3113787
#> [40,] 0.9633028 1.1624264 -0.13249778 0.9949316
#> 
#> $pre[[3]]
#>             [,1]      [,2]
#>  [1,] -0.4071264 0.5235679
#>  [2,] -0.3878778 0.4987872
#>  [3,] -0.4043478 0.5092934
#>  [4,] -0.3694603 0.4694574
#>  [5,] -0.2974783 0.3608997
#>  [6,] -0.3004960 0.3620765
#>  [7,] -0.3822136 0.4831088
#>  [8,] -0.4205758 0.5338017
#>  [9,] -0.3316502 0.3849722
#> [10,] -0.3970196 0.4931160
#> [11,] -0.4072282 0.5131000
#> [12,] -0.3760641 0.4732470
#> [13,] -0.3948351 0.4876077
#> [14,] -0.3464307 0.4361198
#> [15,] -0.4233568 0.5455796
#> [16,] -0.3761184 0.4683385
#> [17,] -0.3223741 0.3776838
#> [18,] -0.4184835 0.5371257
#> [19,] -0.4022106 0.5110587
#> [20,] -0.4101295 0.5170179
#> [21,] -0.4099737 0.5332134
#> [22,] -0.3513659 0.4172634
#> [23,] -0.4164421 0.5384285
#> [24,] -0.4428736 0.5752989
#> [25,] -0.4085554 0.5244063
#> [26,] -0.4077883 0.5143648
#> [27,] -0.3699812 0.4576978
#> [28,] -0.3460702 0.4277062
#> [29,] -0.3865204 0.4853129
#> [30,] -0.3616572 0.4572390
#> [31,] -0.4244530 0.5463989
#> [32,] -0.3685528 0.4598398
#> [33,] -0.4057679 0.5094462
#> [34,] -0.3672308 0.4431926
#> [35,] -0.3547223 0.4501646
#> [36,] -0.3579366 0.4412334
#> [37,] -0.3953720 0.4882079
#> [38,] -0.3998419 0.4865638
#> [39,] -0.3805482 0.4690035
#> [40,] -0.3732477 0.4703201
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.7172161 -0.7172161
#>  [2,]  0.7082015 -0.7082015
#>  [3,]  0.7137447 -0.7137447
#>  [4,]  0.6982372 -0.6982372
#>  [5,]  0.6588959 -0.6588959
#>  [6,]  0.6598380 -0.6598380
#>  [7,]  0.7037715 -0.7037715
#>  [8,]  0.7219947 -0.7219947
#>  [9,]  0.6718628 -0.6718628
#> [10,] -0.2910818  0.2910818
#> [11,]  0.7151090 -0.7151090
#> [12,] -0.2995774  0.2995774
#> [13,]  0.7073282 -0.7073282
#> [14,] -0.3137705  0.3137705
#> [15,]  0.7249075 -0.7249075
#> [16,] -0.3005969  0.3005969
#> [17,]  0.6682006 -0.6682006
#> [18,]  0.7222418 -0.7222418
#> [19,]  0.7136687 -0.7136687
#> [20,]  0.7164962 -0.7164962
#> [21,] -0.2802570  0.2802570
#> [22,]  0.6832243 -0.6832243
#> [23,] -0.2779064  0.2779064
#> [24,] -0.2653835  0.2653835
#> [25,]  0.7176758 -0.7176758
#> [26,]  0.7154806 -0.7154806
#> [27,] -0.3041360  0.3041360
#> [28,] -0.3156628  0.3156628
#> [29,] -0.2948730  0.2948730
#> [30,] -0.3059980  0.3059980
#> [31,] -0.2747107  0.2747107
#> [32,]  0.6960149 -0.6960149
#> [33,]  0.7140660 -0.7140660
#> [34,]  0.6921997 -0.6921997
#> [35,] -0.3089811  0.3089811
#> [36,] -0.3102031  0.3102031
#> [37,]  0.7075635 -0.7075635
#> [38,]  0.7081479 -0.7081479
#> [39,] -0.2995269  0.2995269
#> [40,] -0.3007839  0.3007839
#> 
#> $L
#> [1] 0.6894869 0.7053365 0.7712249 0.6876586 0.7523296 0.8690474
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4057971