Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10         V11         V12          V13         V14
#>  [1,] -0.1607872 -0.14772548 -0.06416137 -0.11330313 -0.053775939 -0.09150165
#>  [2,] -0.3013596 -0.08764049 -0.24543565 -0.15712453 -0.116165624 -0.18168396
#>  [3,] -0.3444452 -0.07485850 -0.20740782 -0.12250008 -0.069904241 -0.12177132
#>  [4,] -0.2734117 -0.09351876 -0.06528071 -0.02553473  0.004683088 -0.05210253
#>  [5,] -0.2529043 -0.14104455 -0.05262820 -0.14538385 -0.105552879 -0.22918963
#>  [6,] -0.2552361 -0.13410268 -0.05989815 -0.03652912 -0.038054444  0.02000142
#>  [7,] -0.3393664 -0.11196935 -0.01502507 -0.07467212 -0.143990180 -0.14357693
#>  [8,] -0.3974010 -0.17272513 -0.07338889 -0.05856934 -0.091786088  0.03140756
#>  [9,] -0.2593134 -0.10234131 -0.06815136 -0.06962684  0.028560518 -0.11265025
#> [10,] -0.3842643 -0.11758339 -0.11277296 -0.10785688 -0.111853368  0.01209965
#>               V15          V16         V17         V18         V19         V2
#>  [1,] -0.22686702 -0.200177808 -0.18426336  0.03326299 -0.05002284 -0.1909429
#>  [2,] -0.19074775 -0.155393533 -0.15978471 -0.11260769  0.04388287 -0.2442241
#>  [3,] -0.07363224  0.008559589  0.06306902 -0.02732932 -0.09789548 -0.3358178
#>  [4,] -0.09391813 -0.066374160 -0.11041329  0.04608821  0.04655490 -0.3509798
#>  [5,] -0.17894558 -0.078056683 -0.22402352 -0.22248958 -0.17746411 -0.1684233
#>  [6,]  0.02397442 -0.025355921 -0.01814099 -0.05760292  0.09299720 -0.2572093
#>  [7,] -0.13519075 -0.048298072 -0.08030830  0.00412420 -0.02343836 -0.3554424
#>  [8,]  0.02057627 -0.016659400  0.09189536  0.03452460  0.07685783 -0.4027570
#>  [9,] -0.07150750  0.061453469  0.08271452  0.16400696  0.14389742 -0.3681345
#> [10,]  0.08216082  0.090698773  0.13412747 -0.01598897  0.07539283 -0.2434606
#>                V20         V21         V22         V23         V24         V25
#>  [1,]  0.001748762 -0.01828212 -0.15632416 -0.07689641  0.01668403 -0.13329480
#>  [2,] -0.049245169 -0.02728145  0.10626450  0.16552723  0.19140362  0.12592013
#>  [3,] -0.056842444 -0.04668848 -0.05898176 -0.04334302 -0.04779599  0.07892657
#>  [4,]  0.052593147 -0.00822119  0.08225163  0.07554404  0.15342965  0.19830280
#>  [5,] -0.003542305  0.01162169  0.11531366  0.08815419  0.03370658  0.12253818
#>  [6,]  0.122537155  0.04435841  0.06301823  0.19517133  0.14051138  0.14161134
#>  [7,] -0.009886170  0.04666999  0.01130648  0.01603364  0.06078855  0.09817323
#>  [8,]  0.089169885  0.08708244  0.10724012  0.18235147  0.16903410  0.02007145
#>  [9,]  0.124789202  0.18896965  0.14046639  0.18947039  0.21890730  0.17520064
#> [10,]  0.077234003  0.22429176  0.22419776  0.22859004  0.16409244  0.12444173
#>              V26         V27         V28         V29         V3         V30
#>  [1,] 0.01738305 -0.03718735  0.03003347 -0.03685740 -0.2980153 0.047867298
#>  [2,] 0.08746533  0.10886425 -0.02412400  0.06646060 -0.2143855 0.024773997
#>  [3,] 0.02516000  0.10830879  0.02044440 -0.01504469 -0.2362171 0.020327298
#>  [4,] 0.18270433  0.16231544  0.18752738  0.08683000 -0.3241945 0.142556712
#>  [5,] 0.02763268  0.10598479  0.22318193  0.12676881 -0.2015752 0.005866976
#>  [6,] 0.15807552  0.20717511  0.24880950  0.24873123 -0.2592496 0.217601152
#>  [7,] 0.17194060  0.13611741  0.25389696  0.08058917 -0.2695990 0.171760857
#>  [8,] 0.18715345  0.16487708  0.09758745  0.14548782 -0.3735277 0.079671362
#>  [9,] 0.11196972  0.21234740  0.17771887  0.17404586 -0.2822711 0.095918692
#> [10,] 0.16902029  0.19675148  0.17407190  0.29428502 -0.3196823 0.150379682
#>                V31          V32          V33         V34          V35
#>  [1,] -0.117894589  0.009141836 -0.154745113 -0.12120971 -0.058229488
#>  [2,] -0.059990104 -0.102062787  0.004476868 -0.16988875 -0.096594484
#>  [3,]  0.057624049 -0.114444174 -0.000223880  0.02502998  0.087580118
#>  [4,]  0.060817840 -0.008068538  0.017066684 -0.02279163 -0.031113109
#>  [5,]  0.008786949 -0.073472023 -0.046984793  0.03966029 -0.092613909
#>  [6,]  0.078026434  0.114187526 -0.016577841  0.07404265 -0.033890476
#>  [7,]  0.190345036  0.018299523  0.065434677  0.09275459  0.008336922
#>  [8,] -0.044750618  0.027069929 -0.050148954 -0.03457162  0.003824998
#>  [9,]  0.168779537 -0.029901543  0.019237024  0.01645866 -0.053531972
#> [10,]  0.055460265  0.126238142 -0.020656218  0.13754534  0.106223489
#>               V36          V37          V38         V39         V4         V40
#>  [1,] -0.12324355 -0.122234993 -0.026313925 -0.03087454 -0.2064117 -0.01895907
#>  [2,] -0.08472199 -0.034375884 -0.183111219 -0.12693195 -0.2586961 -0.08668565
#>  [3,]  0.02963733  0.003247275 -0.068734265 -0.11626702 -0.2675688 -0.02963286
#>  [4,] -0.02067767 -0.143008003  0.013924668 -0.08521486 -0.2852308 -0.02293087
#>  [5,] -0.07429803 -0.032229854 -0.087095969  0.04826762 -0.2312067 -0.02864597
#>  [6,]  0.06099253 -0.037829818  0.051975495 -0.07295484 -0.2559517 -0.01001740
#>  [7,]  0.03947174  0.029565676 -0.006507038 -0.07240392 -0.2736115 -0.09538059
#>  [8,] -0.04562590 -0.024039296 -0.020532055  0.06766001 -0.2817110 -0.02041204
#>  [9,]  0.04172210  0.061617423 -0.084815828 -0.06840963 -0.3004696 -0.01766271
#> [10,]  0.08559239 -0.012511800 -0.018867993  0.01451316 -0.2486555 -0.01689063
#>                V41          V42         V43         V44         V45         V46
#>  [1,] -0.118017025 -0.204449835 -0.21936519 -0.17723332 -0.13442033 -0.09981509
#>  [2,] -0.045336297 -0.169943183 -0.12445231 -0.15521645 -0.11554421 -0.17958796
#>  [3,] -0.078968766 -0.002624159 -0.03114761 -0.15340577 -0.19217760 -0.13873963
#>  [4,] -0.039117920 -0.095552094 -0.05890897 -0.08926750 -0.07190470 -0.21623484
#>  [5,] -0.143815687 -0.104981291 -0.12624021 -0.22900439 -0.24992049 -0.11224464
#>  [6,] -0.111347917 -0.058790278 -0.07890236 -0.05412480 -0.11302108 -0.11278530
#>  [7,] -0.011664772 -0.026402007 -0.12800533 -0.01902517 -0.02642209 -0.15408902
#>  [8,] -0.010659905 -0.040633297 -0.04976027 -0.03118788 -0.19227119 -0.08779788
#>  [9,]  0.028682645  0.002150898 -0.08064972 -0.07778684 -0.14701202 -0.04816905
#> [10,]  0.005996649  0.030247900 -0.04907551 -0.07539259 -0.01909320 -0.17628818
#>               V47        V48        V49         V5        V50        V51
#>  [1,] -0.19460721 -0.1644135 -0.2932928 -0.3207148 -0.2890499 -0.2305757
#>  [2,] -0.15283672 -0.2321708 -0.3625604 -0.2040323 -0.3639810 -0.3795982
#>  [3,] -0.24506926 -0.1711621 -0.3065401 -0.3339295 -0.2985572 -0.4022829
#>  [4,] -0.16490769 -0.2689423 -0.2043953 -0.2001679 -0.2406617 -0.2505920
#>  [5,] -0.13778339 -0.1826516 -0.3219707 -0.1599897 -0.2081747 -0.2284040
#>  [6,] -0.17878782 -0.2436965 -0.1874189 -0.2397704 -0.2161404 -0.2999848
#>  [7,] -0.19428426 -0.1973719 -0.1592573 -0.1720622 -0.2135853 -0.2459081
#>  [8,] -0.27739950 -0.1783603 -0.2804651 -0.2376474 -0.4185768 -0.4515328
#>  [9,] -0.11297779 -0.1771915 -0.3042620 -0.1863748 -0.4507510 -0.3718601
#> [10,] -0.09560302 -0.2445922 -0.2094173 -0.3151914 -0.2899995 -0.2821966
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.2378831 -0.2480537 -0.2658095 -0.2841677 -0.2928115 -0.2642743
#>  [2,] -0.4173825 -0.3327504 -0.3372737 -0.2985183 -0.2649873 -0.2455731
#>  [3,] -0.2959483 -0.3353037 -0.4024176 -0.2854165 -0.3614079 -0.3422818
#>  [4,] -0.3430369 -0.2272683 -0.3846355 -0.3167680 -0.3658421 -0.2911732
#>  [5,] -0.2416444 -0.3276981 -0.2454957 -0.2234801 -0.1989826 -0.3899820
#>  [6,] -0.2355459 -0.3093242 -0.3743247 -0.3543861 -0.3052053 -0.3902653
#>  [7,] -0.3669070 -0.2932101 -0.1848590 -0.3666340 -0.2469010 -0.2697325
#>  [8,] -0.3774343 -0.4443559 -0.3117841 -0.4383749 -0.4585452 -0.3575054
#>  [9,] -0.4317571 -0.4017247 -0.4160289 -0.4022648 -0.4705701 -0.4370854
#> [10,] -0.3734057 -0.3751745 -0.3830129 -0.2717475 -0.3189928 -0.3390350
#>              V58        V59         V6        V60          V7          V8
#>  [1,] -0.2583190 -0.3226121 -0.1924961 -0.3657706 -0.15238746 -0.24960326
#>  [2,] -0.2330263 -0.2451869 -0.3424976 -0.2384482 -0.16984388 -0.18489319
#>  [3,] -0.3726027 -0.2618308 -0.1919567 -0.2666490 -0.17112820 -0.13288061
#>  [4,] -0.3853063 -0.3457242 -0.1872757 -0.4202522 -0.21362600 -0.18292670
#>  [5,] -0.1887371 -0.3394690 -0.1340277 -0.1916307 -0.24790496 -0.19776549
#>  [6,] -0.3580199 -0.4063522 -0.1287678 -0.4008163 -0.06160434 -0.08957886
#>  [7,] -0.2090784 -0.3044674 -0.1238565 -0.3815449 -0.13312854 -0.22836591
#>  [8,] -0.4525148 -0.4650048 -0.1796224 -0.2876022 -0.24056906 -0.14440237
#>  [9,] -0.4620574 -0.4336974 -0.2340145 -0.3984283 -0.22880773 -0.22377156
#> [10,] -0.4117003 -0.3705932 -0.2560921 -0.3543097 -0.20271510 -0.11162621
#>                  V9
#>  [1,] -0.1276921427
#>  [2,] -0.2751004727
#>  [3,] -0.1197922132
#>  [4,] -0.0841639291
#>  [5,] -0.1353363156
#>  [6,]  0.0006996129
#>  [7,] -0.1795672232
#>  [8,] -0.2007556797
#>  [9,] -0.0748859768
#> [10,] -0.1498511781
#> 
#> $W[[2]]
#>             [,1]       [,2]       [,3]       [,4]        [,5]        [,6]
#> [1,] -0.14478862 0.01908558 -0.1142784 0.05956648 -0.03054627 -0.02529657
#> [2,]  0.01791813 0.12870611  0.1933819 0.05425000  0.03103764 -0.07230367
#>             [,7]       [,8]         [,9]      [,10]
#> [1,] -0.03264749 0.01309304  0.062757101 0.01394539
#> [2,]  0.13253695 0.07223596 -0.009515946 0.07822334
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 6.188001e-05 6.877025e-04 7.537668e-04 0.0006239887 0.0004025619
#>  [2,] 6.875149e-05 4.568022e-04 5.198498e-04 0.0005281566 0.0005363255
#>  [3,] 2.261528e-04 2.196482e-03 2.376861e-03 0.0019674429 0.0014855582
#>  [4,] 1.142525e-04 1.531351e-04 2.440545e-04 0.0006145655 0.0008904584
#>  [5,] 5.770496e-05 1.398679e-04 1.755925e-04 0.0002955543 0.0004716148
#>  [6,] 1.065712e-04 9.838335e-05 1.915859e-04 0.0005971195 0.0008381942
#>  [7,] 2.143120e-04 1.351928e-03 1.503639e-03 0.0014712618 0.0014503935
#>  [8,] 1.685207e-04 7.176920e-04 8.321691e-04 0.0010273586 0.0011738779
#>  [9,] 8.756913e-05 4.075748e-06 8.449399e-05 0.0004690969 0.0006769794
#> [10,] 1.372359e-04 6.118907e-04 7.030969e-04 0.0008240852 0.0009098659
#>                V14          V15           V16           V17           V18
#>  [1,] 0.0002664882 8.469758e-05 -0.0001960414 -0.0003329230 -0.0004030241
#>  [2,] 0.0005196101 5.330376e-04  0.0005994037  0.0006675328  0.0007455996
#>  [3,] 0.0012370265 7.429255e-04  0.0000378280 -0.0003233478 -0.0005242740
#>  [4,] 0.0008953909 1.282258e-03  0.0019965445  0.0024358029  0.0027901488
#>  [5,] 0.0005452370 7.372611e-04  0.0010699180  0.0012703205  0.0014469929
#>  [6,] 0.0007916137 1.129814e-03  0.0018318892  0.0022712996  0.0026202084
#>  [7,] 0.0014081509 1.413886e-03  0.0014891499  0.0015634463  0.0016682089
#>  [8,] 0.0011443012 1.355976e-03  0.0017827224  0.0020450093  0.0022799975
#>  [9,] 0.0006193033 9.243113e-04  0.0015242503  0.0018827390  0.0021712270
#> [10,] 0.0008865662 1.022896e-03  0.0012969476  0.0014636368  0.0015975942
#>                 V19           V2           V20           V21           V22
#>  [1,] -0.0002403289 7.667166e-05 -0.0001029117 -0.0001087866 -8.190594e-05
#>  [2,]  0.0009490538 8.757625e-05  0.0010564506  0.0011367792  1.157121e-03
#>  [3,]  0.0001144754 2.697023e-04  0.0007046347  0.0007124258  7.539822e-04
#>  [4,]  0.0031086654 1.511810e-04  0.0031017802  0.0033025640  3.068598e-03
#>  [5,]  0.0016098863 7.672873e-05  0.0016547891  0.0017796456  1.727334e-03
#>  [6,]  0.0029128810 1.418087e-04  0.0028662524  0.0030755975  2.816501e-03
#>  [7,]  0.0022460060 2.612215e-04  0.0026335071  0.0027670948  2.625700e-03
#>  [8,]  0.0027287615 2.111250e-04  0.0029083673  0.0030810899  2.858599e-03
#>  [9,]  0.0024313851 1.142672e-04  0.0023483017  0.0025268814  2.290723e-03
#> [10,]  0.0019369501 1.720779e-04  0.0020918062  0.0021948661  1.972030e-03
#>                 V23           V24           V25           V26           V27
#>  [1,] -0.0005523498 -0.0007556349 -0.0007102992 -0.0005313737 -4.799832e-04
#>  [2,]  0.0011272029  0.0011239102  0.0010875182  0.0011510158  1.205038e-03
#>  [3,] -0.0003952801 -0.0009650993 -0.0008803825 -0.0002791938 -6.566473e-05
#>  [4,]  0.0036091068  0.0039471272  0.0037086369  0.0035555018  3.613969e-03
#>  [5,]  0.0020065573  0.0021681021  0.0020789522  0.0020257493  2.042588e-03
#>  [6,]  0.0033514247  0.0036966534  0.0034761894  0.0033279313  3.367006e-03
#>  [7,]  0.0023373029  0.0022092171  0.0020884934  0.0023415515  2.467682e-03
#>  [8,]  0.0030458240  0.0032009884  0.0030095936  0.0030398073  3.129692e-03
#>  [9,]  0.0028756169  0.0032369877  0.0029759153  0.0028319192  2.918671e-03
#> [10,]  0.0020484512  0.0021310594  0.0019978798  0.0020527401  2.123397e-03
#>                 V28           V29           V3           V30           V31
#>  [1,] -0.0006253083 -0.0007853793 8.385900e-05 -0.0007428836 -0.0009010718
#>  [2,]  0.0012045111  0.0010513400 1.060782e-04  0.0009372081  0.0007509343
#>  [3,] -0.0004032599 -0.0009964449 3.023446e-04 -0.0009858059 -0.0016549704
#>  [4,]  0.0040069069  0.0040141545 1.766314e-04  0.0037057535  0.0035688347
#>  [5,]  0.0021827057  0.0021115909 9.526924e-05  0.0019469907  0.0018204932
#>  [6,]  0.0037245797  0.0037678900 1.617029e-04  0.0034896824  0.0034050810
#>  [7,]  0.0025134136  0.0021566101 3.045162e-04  0.0019871075  0.0014976037
#>  [8,]  0.0034213906  0.0033008580 2.452775e-04  0.0030664968  0.0027877300
#>  [9,]  0.0033681185  0.0035198799 1.305415e-04  0.0032634470  0.0032159992
#> [10,]  0.0023283972  0.0022281426 2.001554e-04  0.0021023547  0.0018941918
#>                 V32           V33           V34           V35           V36
#>  [1,] -0.0007462107 -0.0009495136 -0.0014403557 -0.0016135419 -0.0017040049
#>  [2,]  0.0006936332  0.0006473697  0.0005620624  0.0005183337  0.0004225438
#>  [3,] -0.0013027051 -0.0018261772 -0.0031343179 -0.0036471851 -0.0039296326
#>  [4,]  0.0032472770  0.0035353114  0.0043260634  0.0045771729  0.0043559332
#>  [5,]  0.0016313497  0.0017528398  0.0020518007  0.0021481107  0.0020350290
#>  [6,]  0.0030925924  0.0033728655  0.0041852847  0.0044632884  0.0042863008
#>  [7,]  0.0015009679  0.0014421065  0.0012298592  0.0010751438  0.0007431479
#>  [8,]  0.0026184374  0.0027533874  0.0031264842  0.0032045750  0.0029212001
#>  [9,]  0.0028796544  0.0031335061  0.0039573966  0.0042439587  0.0041247947
#> [10,]  0.0017844419  0.0018623071  0.0021044919  0.0021525883  0.0019349155
#>                 V37           V38           V39           V4           V40
#>  [1,] -0.0012202372 -0.0005873444 -0.0004561502 0.0001251118 -0.0006481011
#>  [2,]  0.0004587924  0.0005431573  0.0005472694 0.0001360058  0.0004621824
#>  [3,] -0.0027213120 -0.0009463772 -0.0006504119 0.0004485875 -0.0013740028
#>  [4,]  0.0036327218  0.0028057628  0.0026082829 0.0001783424  0.0027776851
#>  [5,]  0.0016953937  0.0013005900  0.0011940513 0.0001071869  0.0012504553
#>  [6,]  0.0035825898  0.0027894552  0.0026353765 0.0001526297  0.0028231923
#>  [7,]  0.0009141650  0.0013666369  0.0013805131 0.0003973777  0.0010483604
#>  [8,]  0.0025609339  0.0023012314  0.0021875898 0.0002845999  0.0021211794
#>  [9,]  0.0034052636  0.0026402218  0.0024885172 0.0001070786  0.0026437179
#> [10,]  0.0017331887  0.0016538350  0.0015921586 0.0002350249  0.0014883521
#>                 V41           V42           V43          V44          V45
#>  [1,] -0.0004059383 -0.0001050884 -4.736489e-05 1.131254e-05 0.0004021228
#>  [2,]  0.0004947012  0.0005466915  5.103082e-04 4.804142e-04 0.0004784131
#>  [3,] -0.0006948971  0.0001691242  2.803285e-04 4.024907e-04 0.0015161313
#>  [4,]  0.0024321108  0.0019923872  1.722562e-03 1.556236e-03 0.0007747001
#>  [5,]  0.0011001711  0.0009177915  8.127103e-04 7.264730e-04 0.0003944090
#>  [6,]  0.0024751629  0.0020272321  1.738568e-03 1.557727e-03 0.0007624397
#>  [7,]  0.0011944620  0.0014074137  1.292900e-03 1.236233e-03 0.0013705053
#>  [8,]  0.0020010678  0.0018432991  1.623758e-03 1.507244e-03 0.0011150843
#>  [9,]  0.0022829142  0.0018525822  1.585062e-03 1.392502e-03 0.0006153862
#> [10,]  0.0014208594  0.0013338086  1.167488e-03 1.086980e-03 0.0008618808
#>                V46          V47          V48          V49           V5
#>  [1,] 0.0004878813 0.0002891866 0.0001655863 9.443803e-05 0.0001270269
#>  [2,] 0.0004101009 0.0002914021 0.0002027574 1.100412e-04 0.0001751018
#>  [3,] 0.0016414197 0.0009599414 0.0005944413 3.393073e-04 0.0004628745
#>  [4,] 0.0003962687 0.0003971727 0.0003293372 1.656689e-04 0.0003011966
#>  [5,] 0.0002280043 0.0002127024 0.0001755987 8.929066e-05 0.0001603908
#>  [6,] 0.0003784128 0.0003775703 0.0003102122 1.588205e-04 0.0002802988
#>  [7,] 0.0011792023 0.0007868980 0.0005579075 3.024334e-04 0.0004738430
#>  [8,] 0.0008078978 0.0006114847 0.0004621663 2.414588e-04 0.0003884663
#>  [9,] 0.0002847589 0.0003046509 0.0002533406 1.289058e-04 0.0002349848
#> [10,] 0.0006474448 0.0004759029 0.0003589405 1.902081e-04 0.0003039622
#>                V50          V51          V52           V53          V54
#>  [1,] 9.727035e-06 3.740421e-05 2.492288e-05 -3.554159e-06 1.628999e-06
#>  [2,] 3.950223e-05 3.599415e-05 3.264852e-05  2.202484e-05 2.046677e-05
#>  [3,] 6.411476e-05 1.357720e-04 1.022186e-04  1.327885e-05 3.000685e-05
#>  [4,] 9.372874e-05 3.944717e-05 4.553004e-05  7.200752e-05 5.261698e-05
#>  [5,] 5.162755e-05 2.357000e-05 2.811126e-05  3.668252e-05 2.849138e-05
#>  [6,] 8.703657e-05 3.372170e-05 3.855383e-05  6.758596e-05 4.900346e-05
#>  [7,] 1.030761e-04 1.058783e-04 9.268954e-05  5.899759e-05 5.644099e-05
#>  [8,] 1.028410e-04 7.333299e-05 7.005610e-05  7.000816e-05 5.617421e-05
#>  [9,] 7.395419e-05 2.588251e-05 3.111893e-05  6.254219e-05 4.259018e-05
#> [10,] 7.660063e-05 5.881997e-05 5.534113e-05  5.131158e-05 4.166465e-05
#>                 V55           V56           V57          V58           V59
#>  [1,] -7.652073e-06 -4.144250e-06 -1.046283e-05 3.352924e-06 -4.453996e-06
#>  [2,]  1.509632e-05  1.654857e-05  1.240017e-05 1.657139e-05  1.676919e-05
#>  [3,]  3.151232e-06  4.626530e-06 -1.374199e-05 2.690186e-05  4.933641e-06
#>  [4,]  5.318792e-05  5.590991e-05  4.972970e-05 4.450369e-05  6.288916e-05
#>  [5,]  2.810921e-05  2.897420e-05  2.695234e-05 2.213490e-05  2.950031e-05
#>  [6,]  4.988810e-05  5.258304e-05  4.655642e-05 4.378887e-05  6.064303e-05
#>  [7,]  4.234864e-05  4.313209e-05  2.882054e-05 4.708014e-05  4.689362e-05
#>  [8,]  5.106069e-05  5.223331e-05  4.194219e-05 4.828214e-05  5.966244e-05
#>  [9,]  4.455885e-05  4.652835e-05  4.194921e-05 4.015793e-05  5.694848e-05
#> [10,]  3.835746e-05  3.748231e-05  2.992338e-05 3.784610e-05  4.436406e-05
#>                  V6           V60            V7           V8           V9
#>  [1,] -1.700305e-05 -3.100906e-06 -4.561209e-05 9.614294e-05 5.009967e-04
#>  [2,]  1.776186e-04  1.440567e-05  1.934480e-04 2.281837e-04 3.659179e-04
#>  [3,]  9.169161e-05  5.230987e-06  5.222161e-05 4.690387e-04 1.666861e-03
#>  [4,]  5.709950e-04  5.237342e-05  6.315284e-04 5.115063e-04 1.849294e-04
#>  [5,]  2.812684e-04  2.510872e-05  3.311955e-04 2.713465e-04 1.468827e-04
#>  [6,]  5.384611e-04  5.016140e-05  5.877658e-04 4.838823e-04 1.416550e-04
#>  [7,]  4.375476e-04  3.967189e-05  4.785254e-04 6.288876e-04 1.079939e-03
#>  [8,]  5.255144e-04  4.938646e-05  5.764910e-04 5.887550e-04 6.135915e-04
#>  [9,]  4.727117e-04  4.563583e-05  4.963344e-04 4.104127e-04 6.572817e-05
#> [10,]  3.729285e-04  3.682975e-05  4.136197e-04 4.461117e-04 5.162161e-04
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.03501361  0.05291764  0.06648865  0.07298914  0.05758163  0.08322775
#> [2,] -0.03501361 -0.05291764 -0.06648865 -0.07298914 -0.05758163 -0.08322775
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.07815486  0.07770432  0.08664821  0.08987542
#> [2,] -0.07815486 -0.07770432 -0.08664821 -0.08987542
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.1661507 0.1948377 0.4682103 0.3735901 0.2052961 0.4998313 0.3321175
#>  [8] 0.4949757 0.7303355 0.7050379
#> 
#> $B[[2]]
#> [1] 0.08264798 0.02831730
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] -0.001334059  0.001635277 -0.001853736  0.006661889  0.003454244
#>  [6]  0.006339779  0.003594396  0.005461610  0.005710810  0.003798040
#> 
#> $vB[[2]]
#> [1]  0.09761819 -0.09761819
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0388 0.2718 0.3645 0.3934 0.3843 0.4677 0.5364 0.4823 0.4835 0.5862
#>  [2,] 0.0094 0.1328 0.1303 0.0273 0.0644 0.0712 0.1204 0.0717 0.1224 0.2349
#>  [3,] 0.0209 0.4125 0.3943 0.1334 0.4622 0.9970 0.9137 0.8292 0.6994 0.7825
#>  [4,] 0.0731 0.1114 0.1739 0.3160 0.3249 0.2164 0.2031 0.2580 0.1796 0.2422
#>  [5,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#>  [6,] 0.0239 0.1376 0.0938 0.0259 0.1499 0.2851 0.5743 0.8278 0.8669 0.8131
#>  [7,] 0.0373 0.0802 0.1564 0.2565 0.2624 0.1179 0.0597 0.1563 0.2241 0.3586
#>  [8,] 0.0329 0.2672 0.3056 0.3161 0.2314 0.2067 0.1804 0.2808 0.4423 0.5947
#>  [9,] 0.0152 0.1504 0.1911 0.2115 0.2249 0.2573 0.1701 0.2023 0.2538 0.3417
#> [10,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#> [11,] 0.0270 0.2122 0.2210 0.2399 0.2964 0.4061 0.5095 0.5512 0.6613 0.6804
#> [12,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [13,] 0.0151 0.1205 0.0847 0.1518 0.2305 0.2793 0.3404 0.4527 0.6950 0.8807
#> [14,] 0.0206 0.1573 0.2327 0.1785 0.1507 0.1916 0.2061 0.2307 0.2360 0.1299
#> [15,] 0.0124 0.2120 0.1640 0.1901 0.3026 0.2019 0.0592 0.2390 0.3657 0.3809
#> [16,] 0.0191 0.3134 0.4786 0.5239 0.4393 0.3440 0.2869 0.3889 0.4420 0.3892
#> [17,] 0.0231 0.0347 0.0575 0.1382 0.2274 0.4038 0.5223 0.6847 0.7521 0.7760
#> [18,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#> [19,] 0.0707 0.1247 0.2340 0.1764 0.2284 0.3115 0.4725 0.5543 0.5386 0.3746
#> [20,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [21,] 0.0408 0.2176 0.2459 0.3332 0.3087 0.2613 0.3232 0.3731 0.4203 0.5364
#> [22,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [23,] 0.0089 0.2119 0.3003 0.3094 0.2743 0.2547 0.1870 0.1452 0.1457 0.2429
#> [24,] 0.0298 0.0852 0.2476 0.3645 0.2777 0.2826 0.3237 0.4335 0.5638 0.4555
#> [25,] 0.0229 0.1631 0.2709 0.3358 0.4091 0.4400 0.5485 0.7213 0.8137 0.9185
#> [26,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [27,] 0.0394 0.2558 0.3078 0.3404 0.3400 0.3951 0.3352 0.2252 0.2086 0.2248
#> [28,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [29,] 0.0087 0.0818 0.1315 0.1862 0.2789 0.2579 0.2240 0.2568 0.2933 0.2991
#> [30,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#> [31,] 0.0368 0.0506 0.0906 0.2545 0.1464 0.1272 0.1223 0.1669 0.1424 0.1285
#> [32,] 0.0423 0.2696 0.3412 0.4292 0.3682 0.3940 0.2965 0.3172 0.2825 0.3050
#> [33,] 0.0068 0.2475 0.1606 0.0922 0.2398 0.3220 0.4295 0.2652 0.0666 0.1442
#> [34,] 0.0190 0.0349 0.1459 0.3473 0.3197 0.2823 0.0166 0.0572 0.2164 0.4563
#> [35,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [36,] 0.0131 0.3193 0.3468 0.3738 0.3055 0.1926 0.1385 0.2122 0.2758 0.4576
#> [37,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [38,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [39,] 0.0179 0.5965 0.6254 0.4507 0.3693 0.2864 0.1635 0.0422 0.1785 0.4394
#> [40,] 0.0346 0.1508 0.1809 0.2390 0.2947 0.2866 0.4010 0.5325 0.5486 0.5823
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.7579 0.0324 0.6997 0.6918 0.8633 0.9107 0.9346 0.7884 0.8585 0.9261
#>  [2,] 0.3684 0.0166 0.3918 0.4925 0.8793 0.9606 0.8786 0.6905 0.6937 0.5674
#>  [3,] 0.8789 0.0261 0.8501 0.8920 0.9473 1.0000 0.8975 0.7806 0.8321 0.6502
#>  [4,] 0.3609 0.1249 0.1810 0.2604 0.6572 0.9734 0.9757 0.8079 0.6521 0.4915
#>  [5,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#>  [6,] 0.9045 0.0189 0.9046 1.0000 0.9976 0.9872 0.9761 0.9009 0.9724 0.9675
#>  [7,] 0.1792 0.0281 0.3256 0.6079 0.6988 0.8391 0.8553 0.7710 0.6215 0.5736
#>  [8,] 0.6601 0.0216 0.5844 0.4539 0.4789 0.5646 0.5281 0.7115 1.0000 0.9564
#>  [9,] 0.4026 0.0102 0.4553 0.5525 0.5991 0.5854 0.7114 0.9500 0.9858 1.0000
#> [10,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#> [11,] 0.6520 0.0163 0.6788 0.7811 0.8369 0.8969 0.9856 1.0000 0.9395 0.8917
#> [12,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [13,] 0.9154 0.0320 0.7542 0.6736 0.7146 0.8335 0.7701 0.6993 0.6543 0.5040
#> [14,] 0.3812 0.0132 0.5858 0.4497 0.4876 1.0000 0.8675 0.4718 0.5341 0.6197
#> [15,] 0.5929 0.0433 0.6299 0.5801 0.4574 0.4449 0.3691 0.6446 0.8940 0.8978
#> [16,] 0.4088 0.0173 0.5006 0.7271 0.9385 1.0000 0.9831 0.9932 0.9161 0.8237
#> [17,] 0.7708 0.0351 0.8627 1.0000 0.8873 0.8057 0.8760 0.9066 0.9430 0.8846
#> [18,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#> [19,] 0.4583 0.1252 0.5961 0.7464 0.7644 0.5711 0.6257 0.6695 0.7131 0.7567
#> [20,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [21,] 0.7062 0.0653 0.8196 0.8835 0.8299 0.7609 0.7605 0.8367 0.8905 0.7652
#> [22,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [23,] 0.3259 0.0274 0.3679 0.3355 0.3100 0.3914 0.5280 0.6409 0.7707 0.8754
#> [24,] 0.4348 0.0615 0.6433 0.3932 0.1989 0.3540 0.9165 0.9371 0.4620 0.2771
#> [25,] 1.0000 0.0369 0.9418 0.9116 0.9349 0.7484 0.5146 0.4106 0.3443 0.6981
#> [26,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [27,] 0.3382 0.0420 0.4578 0.6474 0.6708 0.7007 0.7619 0.7745 0.6767 0.7373
#> [28,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [29,] 0.3924 0.0046 0.4691 0.5665 0.6464 0.6774 0.7577 0.8856 0.9419 1.0000
#> [30,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#> [31,] 0.1857 0.0403 0.1136 0.2069 0.0219 0.2400 0.2547 0.0240 0.1923 0.4753
#> [32,] 0.2408 0.0321 0.5420 0.6802 0.6320 0.5824 0.6805 0.5984 0.8412 0.9911
#> [33,] 0.2373 0.0232 0.2595 0.2493 0.3903 0.6384 0.8037 0.7026 0.6874 0.6997
#> [34,] 0.3819 0.0038 0.5627 0.6484 0.7235 0.8242 0.8766 1.0000 0.8582 0.6563
#> [35,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [36,] 0.6487 0.0387 0.7154 0.8010 0.7924 0.8793 1.0000 0.9865 0.9474 0.9474
#> [37,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [38,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [39,] 0.6950 0.0136 0.8097 0.8550 0.8717 0.8601 0.9201 0.8729 0.8084 0.8694
#> [40,] 0.6041 0.0509 0.6749 0.7084 0.7890 0.9284 0.9781 0.9738 1.0000 0.9702
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 0.7080 0.5779 0.0688 0.5215 0.4505 0.3129 0.1448 0.1046 0.1820 0.1519
#>  [2,] 0.6540 0.7802 0.0398 0.7575 0.5836 0.6316 0.8108 0.9039 0.8647 0.6695
#>  [3,] 0.4548 0.4732 0.0120 0.3391 0.2747 0.0978 0.0477 0.1403 0.1834 0.2148
#>  [4,] 0.5363 0.7649 0.1665 0.5250 0.5101 0.4219 0.4160 0.1906 0.0223 0.4219
#>  [5,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#>  [6,] 0.7633 0.4434 0.0466 0.3822 0.4727 0.4007 0.3381 0.3172 0.2222 0.0733
#>  [7,] 0.4402 0.4056 0.0232 0.4411 0.5130 0.5965 0.7272 0.6539 0.5902 0.5393
#>  [8,] 0.6090 0.5112 0.0386 0.4000 0.0482 0.1852 0.2186 0.1436 0.1757 0.1428
#>  [9,] 0.9578 0.8642 0.0113 0.7128 0.5893 0.4323 0.2897 0.1744 0.0770 0.2297
#> [10,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#> [11,] 0.8105 0.6828 0.0341 0.5572 0.4301 0.3339 0.2035 0.0798 0.0809 0.1525
#> [12,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [13,] 0.4926 0.4992 0.0599 0.4161 0.1631 0.0404 0.0637 0.2962 0.3609 0.1866
#> [14,] 0.7143 0.5605 0.0533 0.3728 0.2481 0.1921 0.1386 0.3325 0.2883 0.3228
#> [15,] 0.4980 0.3333 0.0604 0.2350 0.1553 0.3666 0.4340 0.3082 0.3024 0.4109
#> [16,] 0.6957 0.4536 0.0291 0.3281 0.2522 0.3964 0.4154 0.3308 0.1445 0.1923
#> [17,] 0.6500 0.2970 0.0030 0.2423 0.2992 0.2285 0.2277 0.1529 0.1037 0.0352
#> [18,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#> [19,] 0.8077 0.8477 0.1447 0.9289 0.9513 0.7995 0.4362 0.4048 0.4952 0.1712
#> [20,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [21,] 0.5897 0.3037 0.0397 0.0823 0.2787 0.7241 0.8032 0.8050 0.7676 0.7468
#> [22,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [23,] 1.0000 0.9806 0.0248 0.6969 0.4973 0.5020 0.5359 0.3842 0.1848 0.1149
#> [24,] 0.6613 0.8028 0.0650 0.4200 0.5192 0.6962 0.5792 0.8889 0.7863 0.7133
#> [25,] 0.8713 0.9013 0.0040 0.8014 0.4380 0.1319 0.1709 0.2484 0.3044 0.2312
#> [26,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [27,] 0.7834 0.9619 0.0446 1.0000 0.8086 0.5558 0.5409 0.4988 0.3108 0.2897
#> [28,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [29,] 0.8564 0.6790 0.0081 0.5587 0.4147 0.2946 0.2025 0.0688 0.1171 0.2157
#> [30,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#> [31,] 0.7003 0.6825 0.0317 0.6443 0.7063 0.5373 0.6601 0.8708 0.9518 0.9605
#> [32,] 0.9187 0.8005 0.0709 0.6713 0.5632 0.7332 0.6038 0.2575 0.0349 0.1799
#> [33,] 0.8558 1.0000 0.0513 0.9621 0.8996 0.7575 0.6902 0.5686 0.4396 0.4546
#> [34,] 0.5087 0.4817 0.0642 0.4530 0.4521 0.4532 0.5385 0.5308 0.5356 0.5271
#> [35,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [36,] 0.9315 0.8326 0.0329 0.6213 0.3772 0.2822 0.2042 0.2190 0.2223 0.1327
#> [37,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [38,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [39,] 0.8411 0.5793 0.0408 0.3754 0.3485 0.4639 0.6495 0.6901 0.5666 0.5188
#> [40,] 0.9956 0.8235 0.0079 0.6020 0.5342 0.4867 0.3526 0.1566 0.0946 0.1613
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1017 0.1438 0.1986 0.0898 0.2039 0.2778 0.2879 0.1331 0.1140 0.1310
#>  [2,] 0.4027 0.2370 0.2685 0.0359 0.3662 0.3267 0.2200 0.2996 0.2205 0.1163
#>  [3,] 0.1271 0.1912 0.3391 0.0768 0.3444 0.2369 0.1195 0.2665 0.2587 0.1393
#>  [4,] 0.5496 0.2483 0.2034 0.1496 0.2729 0.2837 0.4463 0.3178 0.0807 0.1192
#>  [5,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#>  [6,] 0.2692 0.1888 0.0712 0.0440 0.1062 0.0694 0.0300 0.0893 0.1459 0.1348
#>  [7,] 0.4897 0.4081 0.4145 0.0225 0.6003 0.7196 0.6633 0.6287 0.4087 0.3212
#>  [8,] 0.1644 0.3089 0.3648 0.0627 0.4441 0.3859 0.2813 0.1238 0.0953 0.1201
#>  [9,] 0.2459 0.3101 0.3312 0.0263 0.2220 0.0871 0.2064 0.1808 0.1624 0.1120
#> [10,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#> [11,] 0.2626 0.2456 0.1980 0.0247 0.2412 0.2409 0.1901 0.2077 0.1767 0.1119
#> [12,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [13,] 0.0476 0.1497 0.2405 0.1050 0.1980 0.3175 0.2379 0.1716 0.1559 0.1556
#> [14,] 0.2607 0.2040 0.2396 0.0569 0.1319 0.0683 0.0334 0.0716 0.0976 0.0787
#> [15,] 0.5501 0.4129 0.5499 0.0449 0.5018 0.3132 0.2802 0.2351 0.2298 0.1155
#> [16,] 0.3208 0.3367 0.5683 0.0301 0.5505 0.3231 0.0448 0.3131 0.3387 0.4130
#> [17,] 0.1073 0.1373 0.1331 0.0304 0.1454 0.1115 0.0440 0.0762 0.1381 0.0831
#> [18,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#> [19,] 0.3652 0.3763 0.2841 0.1644 0.0427 0.5331 0.6952 0.4288 0.3063 0.5835
#> [20,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [21,] 0.6253 0.1730 0.2916 0.0604 0.5003 0.5220 0.4824 0.4004 0.3877 0.1651
#> [22,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [23,] 0.1570 0.1311 0.1583 0.0237 0.2631 0.3103 0.4512 0.3785 0.1269 0.1459
#> [24,] 0.7615 0.4401 0.3009 0.0921 0.3163 0.2809 0.2898 0.0526 0.1867 0.1553
#> [25,] 0.1338 0.2056 0.2474 0.0375 0.2790 0.1610 0.0056 0.0351 0.1148 0.1331
#> [26,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [27,] 0.2244 0.0960 0.2287 0.0551 0.3228 0.3454 0.3882 0.3240 0.0926 0.1173
#> [28,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [29,] 0.2216 0.2776 0.2309 0.0230 0.1444 0.1513 0.1745 0.1756 0.1424 0.0908
#> [30,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#> [31,] 0.7712 0.6772 0.6431 0.0293 0.6720 0.6035 0.5155 0.3802 0.2278 0.1522
#> [32,] 0.3039 0.4760 0.5756 0.0108 0.4254 0.5046 0.7179 0.6163 0.5663 0.5749
#> [33,] 0.2959 0.1587 0.1681 0.0444 0.0842 0.1173 0.1754 0.2728 0.1705 0.0194
#> [34,] 0.4260 0.2436 0.1205 0.0452 0.3845 0.4107 0.5067 0.4216 0.2479 0.1586
#> [35,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [36,] 0.0521 0.0618 0.1416 0.0078 0.1460 0.0846 0.1055 0.1639 0.1916 0.2085
#> [37,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [38,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [39,] 0.5060 0.3885 0.3762 0.0633 0.3738 0.2605 0.1591 0.1875 0.2267 0.1577
#> [40,] 0.2824 0.3390 0.3019 0.0243 0.2945 0.2978 0.2676 0.2055 0.2069 0.1625
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.1433 0.0624 0.0100 0.0098 0.1267 0.0131 0.0152 0.0255 0.0071 0.0263
#>  [2,] 0.0635 0.0465 0.0422 0.0174 0.0681 0.0172 0.0134 0.0141 0.0191 0.0145
#>  [3,] 0.1083 0.1383 0.1321 0.1069 0.1064 0.0325 0.0316 0.0057 0.0159 0.0085
#>  [4,] 0.2134 0.3241 0.2945 0.1474 0.1443 0.0211 0.0361 0.0444 0.0230 0.0290
#>  [5,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#>  [6,] 0.0391 0.0546 0.0469 0.0201 0.0657 0.0095 0.0155 0.0091 0.0151 0.0080
#>  [7,] 0.2518 0.1482 0.0988 0.0317 0.0179 0.0269 0.0066 0.0008 0.0045 0.0024
#>  [8,] 0.0825 0.0618 0.0141 0.0108 0.1158 0.0124 0.0104 0.0095 0.0151 0.0059
#>  [9,] 0.0815 0.1117 0.0950 0.0412 0.0097 0.0120 0.0048 0.0049 0.0041 0.0036
#> [10,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#> [11,] 0.0779 0.1344 0.0960 0.0598 0.0822 0.0330 0.0197 0.0189 0.0204 0.0085
#> [12,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [13,] 0.0422 0.0493 0.0476 0.0219 0.1163 0.0059 0.0086 0.0061 0.0015 0.0084
#> [14,] 0.0522 0.0500 0.0231 0.0221 0.0647 0.0144 0.0307 0.0386 0.0147 0.0018
#> [15,] 0.0724 0.0621 0.0318 0.0450 0.0597 0.0167 0.0078 0.0083 0.0057 0.0174
#> [16,] 0.3639 0.2069 0.0859 0.0600 0.0463 0.0267 0.0125 0.0040 0.0136 0.0137
#> [17,] 0.0654 0.0844 0.0595 0.0497 0.0339 0.0313 0.0154 0.0106 0.0097 0.0022
#> [18,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#> [19,] 0.5692 0.2630 0.1196 0.0983 0.1693 0.0374 0.0291 0.0156 0.0197 0.0135
#> [20,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [21,] 0.0442 0.0663 0.0418 0.0475 0.0496 0.0235 0.0066 0.0062 0.0129 0.0184
#> [22,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [23,] 0.1092 0.1485 0.1385 0.0716 0.0224 0.0176 0.0199 0.0096 0.0103 0.0093
#> [24,] 0.1633 0.1252 0.0748 0.0452 0.1615 0.0064 0.0154 0.0031 0.0153 0.0071
#> [25,] 0.0276 0.0763 0.0631 0.0309 0.0455 0.0240 0.0115 0.0064 0.0022 0.0122
#> [26,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [27,] 0.0566 0.0766 0.0969 0.0588 0.0597 0.0050 0.0118 0.0146 0.0040 0.0114
#> [28,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [29,] 0.0138 0.0469 0.0480 0.0159 0.0586 0.0045 0.0015 0.0052 0.0038 0.0079
#> [30,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#> [31,] 0.0801 0.0804 0.0752 0.0566 0.0820 0.0175 0.0058 0.0091 0.0160 0.0160
#> [32,] 0.3593 0.2526 0.2299 0.1271 0.1070 0.0356 0.0367 0.0176 0.0035 0.0093
#> [33,] 0.0213 0.0354 0.0420 0.0093 0.0249 0.0204 0.0199 0.0173 0.0163 0.0055
#> [34,] 0.1124 0.0651 0.0789 0.0325 0.0333 0.0070 0.0026 0.0093 0.0118 0.0112
#> [35,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [36,] 0.2335 0.1964 0.1300 0.0633 0.0721 0.0183 0.0137 0.0150 0.0076 0.0032
#> [37,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [38,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [39,] 0.1211 0.0883 0.0850 0.0355 0.0596 0.0219 0.0086 0.0123 0.0060 0.0187
#> [40,] 0.1216 0.1013 0.0744 0.0386 0.0432 0.0050 0.0146 0.0040 0.0122 0.0107
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0079 0.0111 0.0107 0.0068 0.0097 0.1515 0.0067 0.2134 0.2613 0.2832
#>  [2,] 0.0065 0.0129 0.0217 0.0087 0.0077 0.0706 0.0122 0.1020 0.0893 0.0381
#>  [3,] 0.0372 0.0101 0.0127 0.0288 0.0129 0.1680 0.0023 0.3016 0.3460 0.3314
#>  [4,] 0.0141 0.0161 0.0177 0.0194 0.0207 0.2770 0.0057 0.2555 0.1712 0.0466
#>  [5,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#>  [6,] 0.0018 0.0078 0.0045 0.0026 0.0036 0.0742 0.0024 0.1380 0.1099 0.1384
#>  [7,] 0.0006 0.0073 0.0096 0.0054 0.0085 0.0733 0.0060 0.0841 0.1031 0.0993
#>  [8,] 0.0015 0.0053 0.0016 0.0042 0.0053 0.1482 0.0074 0.2054 0.1605 0.2532
#>  [9,] 0.0013 0.0046 0.0037 0.0011 0.0034 0.0391 0.0033 0.0857 0.0915 0.0949
#> [10,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> [11,] 0.0043 0.0092 0.0138 0.0094 0.0105 0.1256 0.0093 0.1323 0.1584 0.2017
#> [12,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [13,] 0.0128 0.0054 0.0011 0.0019 0.0023 0.1734 0.0062 0.1679 0.1119 0.0889
#> [14,] 0.0100 0.0096 0.0077 0.0180 0.0109 0.1432 0.0070 0.1344 0.2041 0.1571
#> [15,] 0.0188 0.0054 0.0114 0.0196 0.0147 0.0355 0.0062 0.0531 0.0343 0.1052
#> [16,] 0.0172 0.0132 0.0110 0.0122 0.0114 0.0690 0.0068 0.0576 0.1103 0.2423
#> [17,] 0.0052 0.0072 0.0056 0.0038 0.0043 0.0860 0.0030 0.1738 0.1351 0.1063
#> [18,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#> [19,] 0.0127 0.0138 0.0133 0.0131 0.0154 0.0844 0.0218 0.0715 0.0947 0.1583
#> [20,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [21,] 0.0069 0.0198 0.0199 0.0102 0.0070 0.1817 0.0055 0.1178 0.1024 0.0583
#> [22,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [23,] 0.0025 0.0044 0.0021 0.0069 0.0060 0.0845 0.0018 0.1488 0.1224 0.1569
#> [24,] 0.0212 0.0076 0.0152 0.0049 0.0200 0.2294 0.0073 0.2176 0.2033 0.1459
#> [25,] 0.0151 0.0056 0.0026 0.0029 0.0104 0.1452 0.0163 0.2211 0.1188 0.0750
#> [26,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [27,] 0.0032 0.0062 0.0101 0.0068 0.0053 0.1416 0.0087 0.0956 0.0802 0.1618
#> [28,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [29,] 0.0114 0.0050 0.0030 0.0064 0.0058 0.0682 0.0030 0.0993 0.0717 0.0576
#> [30,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#> [31,] 0.0081 0.0070 0.0135 0.0067 0.0078 0.1342 0.0068 0.1161 0.0663 0.0155
#> [32,] 0.0121 0.0075 0.0056 0.0021 0.0043 0.0973 0.0017 0.0961 0.1323 0.2462
#> [33,] 0.0045 0.0068 0.0041 0.0052 0.0194 0.0637 0.0105 0.0422 0.1130 0.1911
#> [34,] 0.0094 0.0140 0.0072 0.0022 0.0055 0.0690 0.0122 0.0901 0.1454 0.0740
#> [35,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [36,] 0.0037 0.0071 0.0040 0.0009 0.0015 0.1341 0.0085 0.1626 0.1902 0.2610
#> [37,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [38,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [39,] 0.0111 0.0126 0.0081 0.0155 0.0160 0.0808 0.0085 0.2090 0.3465 0.5276
#> [40,] 0.0112 0.0102 0.0052 0.0024 0.0079 0.0735 0.0031 0.0938 0.1134 0.1228
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2206377 0.4064970 0.4285405 0.7091510 0.4082344 0.8347242 0.6303438
#>  [2,] 0.2466950 0.4730987 0.5186662 0.7298950 0.5244130 0.8517491 0.7394501
#>  [3,] 0.1647712 0.3078685 0.3659209 0.6352872 0.2915318 0.7956770 0.5185910
#>  [4,] 0.1974145 0.3857905 0.3904862 0.6123994 0.4154033 0.7644633 0.5894477
#>  [5,] 0.1937998 0.3319096 0.3783745 0.7013027 0.4158222 0.8404351 0.6527021
#>  [6,] 0.2328518 0.4752474 0.5011751 0.7540540 0.4488321 0.8673293 0.6979255
#>  [7,] 0.2010810 0.3743080 0.4637002 0.6575868 0.4128938 0.7801790 0.6583715
#>  [8,] 0.2861303 0.4225436 0.4752608 0.6969019 0.4362594 0.7991522 0.6356874
#>  [9,] 0.3062271 0.5025302 0.5257921 0.7717948 0.5547500 0.8680395 0.7418724
#> [10,] 0.2760166 0.4674935 0.5343129 0.7470244 0.5424242 0.8650992 0.7598850
#> [11,] 0.2252405 0.4371341 0.4588716 0.7363385 0.4365084 0.8519292 0.6674917
#> [12,] 0.1816688 0.2645402 0.4459832 0.6429917 0.3661378 0.8127976 0.7039157
#> [13,] 0.2649877 0.4344047 0.4768625 0.6883823 0.3986816 0.7987256 0.6194316
#> [14,] 0.3158844 0.4819972 0.4952764 0.6913316 0.5231885 0.8164127 0.6553167
#> [15,] 0.2795149 0.4110182 0.4945760 0.6689757 0.4444656 0.7791840 0.6447375
#> [16,] 0.2064953 0.3902053 0.3998895 0.6949953 0.4264256 0.8120820 0.6264783
#> [17,] 0.2631806 0.4740287 0.4973338 0.7352506 0.4435320 0.8386864 0.6595072
#> [18,] 0.3245932 0.5329080 0.5278862 0.7301249 0.5574587 0.8235098 0.6994644
#> [19,] 0.1522737 0.2790027 0.3790254 0.6223320 0.3414289 0.8008663 0.6451871
#> [20,] 0.2690361 0.3943413 0.5291009 0.7041033 0.4806109 0.8467353 0.7473652
#> [21,] 0.1732408 0.3439608 0.4460549 0.6644658 0.3712749 0.8048338 0.6534774
#> [22,] 0.3081438 0.5157735 0.5087970 0.7606185 0.5602379 0.8544683 0.7324840
#> [23,] 0.2804303 0.4191005 0.4961313 0.7147278 0.4977383 0.8296731 0.7023870
#> [24,] 0.1962098 0.2924273 0.4484805 0.6197611 0.3687527 0.7935761 0.6420584
#> [25,] 0.2436548 0.3762293 0.4556507 0.7007923 0.3963035 0.8428351 0.6457234
#> [26,] 0.2977482 0.4609473 0.5046064 0.7327750 0.5301749 0.8399951 0.7094819
#> [27,] 0.2358891 0.4145761 0.4770038 0.7266561 0.4870721 0.8520483 0.7093423
#> [28,] 0.1548887 0.2460471 0.3785663 0.5508350 0.2986678 0.7220071 0.5946582
#> [29,] 0.3180331 0.5299750 0.5384114 0.7670509 0.5538954 0.8569218 0.7247349
#> [30,] 0.1761597 0.1912597 0.4146297 0.5400085 0.2754311 0.7328453 0.5941592
#> [31,] 0.2426110 0.2811378 0.5023886 0.5756917 0.4067864 0.7451557 0.6633814
#> [32,] 0.1713800 0.2884007 0.3649472 0.6479492 0.3652433 0.8004322 0.6364950
#> [33,] 0.2695526 0.4525771 0.5316878 0.7419134 0.5281031 0.8647637 0.7423895
#> [34,] 0.2393329 0.4484271 0.4986752 0.7186695 0.4585795 0.8229911 0.6896379
#> [35,] 0.2637965 0.4825952 0.4950175 0.7645865 0.5421573 0.8722780 0.7254292
#> [36,] 0.2662149 0.4833906 0.4562502 0.7578007 0.5057220 0.8612184 0.6970495
#> [37,] 0.1977483 0.3836212 0.4205443 0.7067237 0.4066565 0.8291832 0.6411980
#> [38,] 0.2467677 0.2989452 0.5306873 0.6130772 0.4391989 0.7793379 0.6940539
#> [39,] 0.1978989 0.3569922 0.4060528 0.6982191 0.4233559 0.8356714 0.6576864
#> [40,] 0.2381880 0.4616350 0.4773237 0.7633409 0.4836701 0.8731170 0.7136436
#>            [,8]      [,9]     [,10]
#>  [1,] 0.7625470 0.8801520 0.8907604
#>  [2,] 0.7611911 0.8772444 0.9150649
#>  [3,] 0.7404351 0.8550232 0.8820083
#>  [4,] 0.6445926 0.8080511 0.8240770
#>  [5,] 0.7387726 0.8673894 0.8942556
#>  [6,] 0.8219125 0.9223815 0.9312737
#>  [7,] 0.7063451 0.8460320 0.8752895
#>  [8,] 0.7407266 0.8480404 0.8542800
#>  [9,] 0.7876832 0.8921919 0.9054958
#> [10,] 0.7729646 0.8855909 0.9127461
#> [11,] 0.7884560 0.8999990 0.9095113
#> [12,] 0.6485832 0.8125220 0.8594824
#> [13,] 0.7641922 0.8730024 0.8818751
#> [14,] 0.7409554 0.8423939 0.8653277
#> [15,] 0.7266607 0.8382303 0.8530501
#> [16,] 0.7443343 0.8669751 0.8835331
#> [17,] 0.8026263 0.9020723 0.9088106
#> [18,] 0.7588779 0.8607588 0.8782433
#> [19,] 0.6800184 0.8476973 0.8876087
#> [20,] 0.7287732 0.8638539 0.8950622
#> [21,] 0.7365885 0.8705245 0.8993452
#> [22,] 0.7616384 0.8809915 0.8891249
#> [23,] 0.7206686 0.8462367 0.8687043
#> [24,] 0.6698782 0.8292651 0.8634514
#> [25,] 0.7811251 0.8956055 0.9093145
#> [26,] 0.7455245 0.8639951 0.8788457
#> [27,] 0.7459181 0.8758530 0.9012491
#> [28,] 0.6237902 0.7774865 0.8357140
#> [29,] 0.7928978 0.8886300 0.9003215
#> [30,] 0.6240395 0.7836538 0.8275416
#> [31,] 0.6307779 0.7625001 0.8269421
#> [32,] 0.6923899 0.8342092 0.8680719
#> [33,] 0.7462672 0.8676108 0.9008224
#> [34,] 0.7486163 0.8750215 0.8903590
#> [35,] 0.7928762 0.8979050 0.9157562
#> [36,] 0.7752360 0.8969426 0.9027777
#> [37,] 0.7409995 0.8735462 0.8825429
#> [38,] 0.6569270 0.7936198 0.8571705
#> [39,] 0.7232199 0.8709497 0.8880470
#> [40,] 0.8072490 0.9117329 0.9229553
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.6999117 0.3000883
#>  [2,] 0.6989826 0.3010174
#>  [3,] 0.6962038 0.3037962
#>  [4,] 0.6878540 0.3121460
#>  [5,] 0.7013282 0.2986718
#>  [6,] 0.7045146 0.2954854
#>  [7,] 0.6916086 0.3083914
#>  [8,] 0.6921936 0.3078064
#>  [9,] 0.7010119 0.2989881
#> [10,] 0.6997628 0.3002372
#> [11,] 0.7025647 0.2974353
#> [12,] 0.6898013 0.3101987
#> [13,] 0.6943118 0.3056882
#> [14,] 0.6923549 0.3076451
#> [15,] 0.6890419 0.3109581
#> [16,] 0.6987737 0.3012263
#> [17,] 0.7002344 0.2997656
#> [18,] 0.6944508 0.3055492
#> [19,] 0.6946592 0.3053408
#> [20,] 0.6948207 0.3051793
#> [21,] 0.6961961 0.3038039
#> [22,] 0.6992025 0.3007975
#> [23,] 0.6939170 0.3060830
#> [24,] 0.6893307 0.3106693
#> [25,] 0.7000508 0.2999492
#> [26,] 0.6963096 0.3036904
#> [27,] 0.6995996 0.3004004
#> [28,] 0.6821815 0.3178185
#> [29,] 0.6994675 0.3005325
#> [30,] 0.6804372 0.3195628
#> [31,] 0.6778235 0.3221765
#> [32,] 0.6949586 0.3050414
#> [33,] 0.6979307 0.3020693
#> [34,] 0.6966074 0.3033926
#> [35,] 0.7034702 0.2965298
#> [36,] 0.7030578 0.2969422
#> [37,] 0.6993292 0.3006708
#> [38,] 0.6827544 0.3172456
#> [39,] 0.6996974 0.3003026
#> [40,] 0.7051603 0.2948397
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>             [,1]        [,2]         [,3]      [,4]         [,5]      [,6]
#>  [1,] -1.2619540 -0.37846565 -0.287808292 0.8912641 -0.371269223 1.6194858
#>  [2,] -1.1163177 -0.10770932  0.074699541 0.9940897  0.097729664 1.7483856
#>  [3,] -1.6231478 -0.81010308 -0.549756446 0.5549668 -0.887955960 1.3594913
#>  [4,] -1.4025326 -0.46504171 -0.445268879 0.4574091 -0.341672152 1.1773075
#>  [5,] -1.4255064 -0.69956107 -0.496453563 0.8535091 -0.339947603 1.6614689
#>  [6,] -1.1922780 -0.09909152  0.004700368 1.1203519 -0.205390676 1.8775489
#>  [7,] -1.3795520 -0.51377942 -0.145455093 0.6525587 -0.352015358 1.2667097
#>  [8,] -0.9142530 -0.31234011 -0.099037838 0.8325881 -0.256357359 1.3810039
#>  [9,] -0.8178176  0.01012105  0.103260053 1.2184734  0.219881644 1.8837347
#> [10,] -0.9643076 -0.13020964  0.137467796 1.0828049  0.170105994 1.8583041
#> [11,] -1.2353840 -0.25280134 -0.164886328 1.0270238 -0.255344622 1.7498126
#> [12,] -1.5050821 -1.02250259 -0.216913600 0.5883730 -0.548821824 1.4682918
#> [13,] -1.0202037 -0.26390212 -0.092616317 0.7925672 -0.410961332 1.3783481
#> [14,] -0.7727506 -0.07204222 -0.018894825 0.8063518  0.092820411 1.4922296
#> [15,] -0.9468689 -0.35975782 -0.021696687 0.7035561 -0.223057758 1.2609173
#> [16,] -1.3461818 -0.44644928 -0.405925633 0.8235778 -0.296449600 1.4635957
#> [17,] -1.0295025 -0.10397882 -0.010664991 1.0214279 -0.226839663 1.6484869
#> [18,] -0.7327423  0.13182253  0.111660862 0.9952561  0.230854556 1.5403103
#> [19,] -1.7168780 -0.94941389 -0.493686896 0.4994582 -0.656933042 1.3917175
#> [20,] -0.9995182 -0.42909980  0.116535479 0.8669147 -0.077595471 1.7092216
#> [21,] -1.5628312 -0.64569271 -0.216623555 0.6832596 -0.526751504 1.4167840
#> [22,] -0.8088114  0.06311499  0.035191604 1.1560735  0.242127435 1.7700854
#> [23,] -0.9423282 -0.32646717 -0.015475301 0.9184578 -0.009047059 1.5833121
#> [24,] -1.4101539 -0.88362441 -0.206812153 0.4885343 -0.537571391 1.3466177
#> [25,] -1.1327452 -0.50558404 -0.177864673 0.8510737 -0.420891264 1.6794757
#> [26,] -0.8580440 -0.15652981  0.018426185 1.0087476  0.120846560 1.6581913
#> [27,] -1.1753514 -0.34507954 -0.092049901 0.9777224 -0.051722976 1.7507574
#> [28,] -1.6967619 -1.11980677 -0.495637939 0.2040451 -0.853649941 0.9544393
#> [29,] -0.7628256  0.12004391  0.153949066 1.1917333  0.216422578 1.7899553
#> [30,] -1.5425859 -1.44184550 -0.344858854 0.1603771 -0.967239420 1.0091069
#> [31,] -1.1384178 -0.93882497  0.009554479 0.3051118 -0.377266167 1.0729407
#> [32,] -1.5758783 -0.90316425 -0.553955259 0.6100366 -0.552677974 1.3889981
#> [33,] -0.9968937 -0.19026344  0.126921479 1.0559372  0.112530903 1.8554330
#> [34,] -1.1563404 -0.20702788 -0.005299040 0.9378713 -0.166062403 1.5367455
#> [35,] -1.0263285 -0.06964740 -0.019930641 1.1779915  0.169030638 1.9212523
#> [36,] -1.0139124 -0.06646204 -0.175447936 1.1406594  0.022888980 1.8254469
#> [37,] -1.4004274 -0.47420628 -0.320539470 0.8795245 -0.377804559 1.5798493
#> [38,] -1.1159261 -0.85232596  0.122903596 0.4602655 -0.244413977 1.2618123
#> [39,] -1.3994784 -0.58844282 -0.380307368 0.8388319 -0.309011974 1.6263676
#> [40,] -1.1626397 -0.15376212 -0.090767483 1.1710839 -0.065342959 1.9288041
#>             [,7]      [,8]     [,9]    [,10]
#>  [1,] 0.53369191 1.1666943 1.993870 2.098532
#>  [2,] 1.04311234 1.1592209 1.966590 2.377107
#>  [3,] 0.07439843 1.0482311 1.774555 2.011587
#>  [4,] 0.36168271 0.5953536 1.437396 1.544217
#>  [5,] 0.63093777 1.0395989 1.878071 2.134967
#>  [6,] 0.83743878 1.5293591 2.475153 2.606421
#>  [7,] 0.65604538 0.8776987 1.703813 1.948559
#>  [8,] 0.55669451 1.0497485 1.719313 1.768572
#>  [9,] 1.05572332 1.3110167 2.113329 2.259839
#> [10,] 1.15204891 1.2251272 2.046475 2.347635
#> [11,] 0.69686201 1.3156434 2.197213 2.307682
#> [12,] 0.86601418 0.6128174 1.466482 1.810997
#> [13,] 0.48713648 1.1758023 1.927770 2.010308
#> [14,] 0.64249253 1.0509400 1.676149 1.860264
#> [15,] 0.59598637 0.9777459 1.645119 1.758726
#> [16,] 0.51713819 1.0686197 1.874474 2.026322
#> [17,] 0.66109888 1.4027900 2.220465 2.299198
#> [18,] 0.84474849 1.1465374 1.821607 1.975899
#> [19,] 0.59794958 0.7538562 1.716654 2.066544
#> [20,] 1.08460924 0.9884072 1.847675 2.143526
#> [21,] 0.63435978 1.0283120 1.905604 2.189970
#> [22,] 1.00726204 1.1616828 2.001853 2.081834
#> [23,] 0.85869075 0.9477804 1.705385 1.889551
#> [24,] 0.58430937 0.7076343 1.580428 1.844257
#> [25,] 0.60029346 1.2722350 2.149323 2.305294
#> [26,] 0.89286916 1.0748834 1.848876 1.981545
#> [27,] 0.89219207 1.0769592 1.953732 2.211181
#> [28,] 0.38325605 0.5056671 1.251078 1.626678
#> [29,] 0.96807114 1.3424820 2.076823 2.200802
#> [30,] 0.38118628 0.5067295 1.287087 1.568303
#> [31,] 0.67839939 0.5355556 1.166435 1.564108
#> [32,] 0.56018338 0.8113164 1.615758 1.884018
#> [33,] 1.05842518 1.0788019 1.879997 2.206395
#> [34,] 0.79842701 1.0912461 1.946107 2.094413
#> [35,] 0.97155438 1.3423506 2.174160 2.386035
#> [36,] 0.83328690 1.2381166 2.163706 2.228476
#> [37,] 0.58056748 1.0511699 1.932684 2.016735
#> [38,] 0.81914055 0.6496297 1.346885 1.791985
#> [39,] 0.65300093 0.9604899 1.909382 2.070945
#> [40,] 0.91314635 1.4322327 2.334979 2.483195
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.6401316 -0.2067458
#>  [2,] 0.6515675 -0.1908905
#>  [3,] 0.6151908 -0.2140946
#>  [4,] 0.5960407 -0.1940648
#>  [5,] 0.6369151 -0.2167156
#>  [6,] 0.6681895 -0.2006999
#>  [7,] 0.6178800 -0.1897705
#>  [8,] 0.6186149 -0.1917796
#>  [9,] 0.6578843 -0.1942365
#> [10,] 0.6537335 -0.1924352
#> [11,] 0.6547027 -0.2048381
#> [12,] 0.6018252 -0.1973652
#> [13,] 0.6275286 -0.1928268
#> [14,] 0.6225021 -0.1886496
#> [15,] 0.6104928 -0.1851510
#> [16,] 0.6346353 -0.2068296
#> [17,] 0.6517950 -0.1966193
#> [18,] 0.6378544 -0.1831562
#> [19,] 0.6135526 -0.2084404
#> [20,] 0.6313914 -0.1913629
#> [21,] 0.6294894 -0.1997594
#> [22,] 0.6511552 -0.1923480
#> [23,] 0.6262306 -0.1922653
#> [24,] 0.6000275 -0.1969647
#> [25,] 0.6397911 -0.2077487
#> [26,] 0.6370906 -0.1926950
#> [27,] 0.6447103 -0.2006814
#> [28,] 0.5724304 -0.1913850
#> [29,] 0.6552896 -0.1894740
#> [30,] 0.5630645 -0.1927171
#> [31,] 0.5675810 -0.1762066
#> [32,] 0.6135034 -0.2099015
#> [33,] 0.6452659 -0.1921977
#> [34,] 0.6403151 -0.1908790
#> [35,] 0.6631691 -0.2007086
#> [36,] 0.6579039 -0.2039976
#> [37,] 0.6373913 -0.2067146
#> [38,] 0.5864245 -0.1800343
#> [39,] 0.6351694 -0.2106878
#> [40,] 0.6672009 -0.2047923
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.3000883 -0.3000883
#>  [2,] -0.6989826  0.6989826
#>  [3,]  0.3037962 -0.3037962
#>  [4,]  0.3121460 -0.3121460
#>  [5,]  0.2986718 -0.2986718
#>  [6,] -0.7045146  0.7045146
#>  [7,] -0.6916086  0.6916086
#>  [8,]  0.3078064 -0.3078064
#>  [9,] -0.7010119  0.7010119
#> [10,] -0.6997628  0.6997628
#> [11,]  0.2974353 -0.2974353
#> [12,]  0.3101987 -0.3101987
#> [13,] -0.6943118  0.6943118
#> [14,] -0.6923549  0.6923549
#> [15,] -0.6890419  0.6890419
#> [16,]  0.3012263 -0.3012263
#> [17,] -0.7002344  0.7002344
#> [18,] -0.6944508  0.6944508
#> [19,]  0.3053408 -0.3053408
#> [20,] -0.6948207  0.6948207
#> [21,] -0.6961961  0.6961961
#> [22,]  0.3007975 -0.3007975
#> [23,]  0.3060830 -0.3060830
#> [24,] -0.6893307  0.6893307
#> [25,] -0.7000508  0.7000508
#> [26,]  0.3036904 -0.3036904
#> [27,]  0.3004004 -0.3004004
#> [28,] -0.6821815  0.6821815
#> [29,] -0.6994675  0.6994675
#> [30,] -0.6804372  0.6804372
#> [31,] -0.6778235  0.6778235
#> [32,]  0.3050414 -0.3050414
#> [33,] -0.6979307  0.6979307
#> [34,] -0.6966074  0.6966074
#> [35,]  0.2965298 -0.2965298
#> [36,]  0.2969422 -0.2969422
#> [37,]  0.3006708 -0.3006708
#> [38,] -0.6827544  0.6827544
#> [39,]  0.3003026 -0.3003026
#> [40,]  0.2948397 -0.2948397
#> 
#> $L
#> [1] 0.6952061 0.6814003 0.6830175 0.7038174 0.7879727 0.7919427
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.5072464