Deep Neural Network with DBN Pretraining Learner

Deep neural network with weights initialized by a deep belief network. Calls deepnet::dbn.dnn.train() from deepnet.

This learner works with tasks that have at least two features.

Initial parameter values

output is set to "softmax" to enable class probabilities.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.dbnDNN")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, deepnet

Parameters

Id	Type	Default	Levels	Range

hidden	untyped	-		-

activationfun	character	-	sigm, linear, tanh	-

learningrate	numeric	-		\([0, \infty)\)

momentum	numeric	-		\([0, \infty)\)

learningrate_scale	numeric	-		\([0, \infty)\)

numepochs	integer	-		\([1, \infty)\)

batchsize	integer	-		\([1, \infty)\)

output	character	-	sigm, linear, softmax	-

hidden_dropout	numeric	-		\([0, 1]\)

visible_dropout	numeric	-		\([0, 1]\)

cd	integer	-		\([1, \infty)\)

References

Rong X (2014). deepnet: Deep Learning Toolkit in R. R package version 0.2.1, https://CRAN.R-project.org/package=deepnet.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifDbnDNN

Methods

Public methods

LearnerClassifDbnDNN$new()

LearnerClassifDbnDNN$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifDbnDNN$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifDbnDNN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.dbnDNN")
print(learner)
#> 
#> ── <LearnerClassifDbnDNN> (classif.dbnDNN): Deep Belief Network DNN ────────────
#> • Model: -
#> • Parameters: hidden=10, activationfun=sigm, learningrate=0.8, momentum=0.5,
#> learningrate_scale=1, numepochs=3, batchsize=100, output=softmax,
#> hidden_dropout=0, visible_dropout=0, cd=1
#> • Packages: mlr3 and deepnet
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> begin to train dbn ......
#> training layer 1 rbm ...
#> dbn has been trained.
#> begin to train deep nn ......
#> deep nn has been trained.

print(learner$model)
#> $input_dim
#> [1] 60
#> 
#> $output_dim
#> [1] 2
#> 
#> $hidden
#> [1] 10
#> 
#> $size
#> [1] 60 10  2
#> 
#> $activationfun
#> [1] "sigm"
#> 
#> $learningrate
#> [1] 0.8
#> 
#> $momentum
#> [1] 0.5
#> 
#> $learningrate_scale
#> [1] 1
#> 
#> $hidden_dropout
#> [1] 0
#> 
#> $visible_dropout
#> [1] 0
#> 
#> $output
#> [1] "softmax"
#> 
#> $W
#> $W[[1]]
#>               V1         V10          V11          V12         V13          V14
#>  [1,] -0.2096996 -0.17777430 -0.074529451 -0.068136587 -0.03058504  0.018485398
#>  [2,] -0.2714864 -0.03491016 -0.073216342 -0.165377446 -0.03969524 -0.109927134
#>  [3,] -0.3286520 -0.01845994 -0.030984997  0.008332851  0.05159431  0.022126661
#>  [4,] -0.2322922 -0.17773591 -0.005784001 -0.037171027 -0.05104788 -0.041777878
#>  [5,] -0.2433296 -0.07900030 -0.172086316 -0.088222814 -0.09688398 -0.073252036
#>  [6,] -0.3655745 -0.03942912 -0.108055747 -0.077382354 -0.09046380 -0.125968281
#>  [7,] -0.3279516 -0.10686523 -0.112325627 -0.073244122 -0.13045754 -0.100314285
#>  [8,] -0.3000296 -0.16339711 -0.147060551 -0.132082449 -0.14958666 -0.159148707
#>  [9,] -0.2751057 -0.08443747 -0.056292746 -0.148984285 -0.16711939 -0.096633081
#> [10,] -0.2085636 -0.12739247 -0.049107479  0.007430761 -0.04998673  0.003429898
#>               V15          V16          V17         V18         V19         V2
#>  [1,] -0.02072977 -0.099192319  0.024073160 -0.03484289  0.03862434 -0.2028613
#>  [2,] -0.02795477  0.033898506 -0.041573656 -0.04567320 -0.02993514 -0.2316938
#>  [3,] -0.09206310  0.068918135 -0.017134709  0.04777615  0.01700415 -0.3511072
#>  [4,] -0.04842212 -0.035579126 -0.042195452  0.09580703  0.07341637 -0.2307114
#>  [5,]  0.00622725 -0.020045195  0.004859359  0.09540660 -0.01593948 -0.3777170
#>  [6,] -0.01765243  0.004744901  0.018432831  0.03763723  0.07549885 -0.3150073
#>  [7,] -0.11954899  0.008429874 -0.031507238 -0.02970750 -0.07362916 -0.2742359
#>  [8,] -0.05302685  0.038438043 -0.052974572 -0.11578809  0.06972287 -0.3386580
#>  [9,]  0.02330542 -0.046912964 -0.031654550 -0.00620943  0.03679207 -0.3727908
#> [10,] -0.04557154 -0.065507995 -0.067690537 -0.01810262  0.06310779 -0.2081745
#>                V20        V21         V22         V23        V24         V25
#>  [1,]  0.123271026 0.10501942  0.09843233  0.17424540 0.09400543  0.08789423
#>  [2,] -0.090463293 0.07790965 -0.01136504 -0.05544767 0.14894725  0.11322960
#>  [3,]  0.201460132 0.18919179  0.08888175  0.14347607 0.16129718  0.29174125
#>  [4,]  0.006331154 0.03607563  0.23175197  0.14624506 0.06346952  0.11124556
#>  [5,]  0.179577414 0.22217383  0.23452772  0.23388168 0.24803976  0.27480133
#>  [6,]  0.060671234 0.08785896  0.01195492  0.15458973 0.05149070  0.05797345
#>  [7,]  0.060590650 0.00487149  0.01699604  0.05240975 0.07142148 -0.02447776
#>  [8,] -0.012860793 0.08439533  0.07427340  0.03749454 0.08002669  0.16629984
#>  [9,]  0.037838266 0.03017810  0.18675999  0.22595661 0.20177622  0.22457654
#> [10,]  0.009807379 0.09279418  0.10188077  0.16402924 0.20281855  0.13937036
#>              V26        V27        V28           V29         V3          V30
#>  [1,] 0.09102350 0.24265885 0.23125458  0.0272620308 -0.1596013  0.052936681
#>  [2,] 0.15044643 0.19161704 0.07294325  0.1264491949 -0.2988004  0.119017122
#>  [3,] 0.27600292 0.24176820 0.27345049  0.1292140580 -0.3669157  0.087866125
#>  [4,] 0.24642510 0.13497766 0.22200769  0.1722085048 -0.3212744 -0.001563728
#>  [5,] 0.20930852 0.25963662 0.07378042  0.1266741397 -0.3091361  0.002185437
#>  [6,] 0.19988132 0.20269170 0.07603569  0.0887247852 -0.3074753  0.014757518
#>  [7,] 0.04063958 0.07364037 0.21865756  0.0773924646 -0.2497507  0.141111217
#>  [8,] 0.08173532 0.03940942 0.08459673 -0.0002509056 -0.1840345  0.022958850
#>  [9,] 0.29689563 0.29614954 0.09389394  0.1438481767 -0.3097763  0.155605221
#> [10,] 0.18274213 0.27831041 0.29366051  0.2181832400 -0.2131444  0.086038844
#>               V31         V32          V33          V34         V35
#>  [1,]  0.14014392 -0.04198069  0.003463547  0.062366930 -0.10904948
#>  [2,]  0.04814325 -0.05121532 -0.037306220 -0.112013439 -0.06529281
#>  [3,] -0.02688891 -0.05269062 -0.010379751 -0.064724314  0.11833063
#>  [4,] -0.02507831  0.04872530  0.074244893 -0.016074553  0.06222207
#>  [5,] -0.07395996  0.06552248 -0.030786195 -0.163529828 -0.18071137
#>  [6,]  0.05242780 -0.02408430 -0.018372405 -0.008617093 -0.12486525
#>  [7,]  0.13122908  0.13287652 -0.048779398 -0.040583141 -0.05365121
#>  [8,]  0.02088100 -0.01512797  0.031541375 -0.023513758  0.08693039
#>  [9,]  0.06478852  0.06207023  0.014726268 -0.007367528 -0.11431513
#> [10,]  0.18478929  0.01168589  0.006922049 -0.013526987  0.10371960
#>                V36          V37         V38         V39         V4         V40
#>  [1,] -0.015817408 -0.018117886  0.01683625  0.07870024 -0.2973028  0.07683130
#>  [2,] -0.037503230 -0.039590019 -0.01451978 -0.19879714 -0.2821932  0.01540094
#>  [3,]  0.161474170  0.040957211  0.03437907 -0.03399065 -0.3350781 -0.08688411
#>  [4,] -0.099610645  0.118275572 -0.10665150 -0.02162132 -0.1734598 -0.12677586
#>  [5,] -0.170548911 -0.095263214 -0.15743326 -0.14937256 -0.3323256 -0.08287587
#>  [6,]  0.001621783  0.016569425  0.02113147 -0.13447817 -0.3102742 -0.04711910
#>  [7,] -0.009846272  0.008228228  0.08132747  0.05508510 -0.1384629 -0.10844416
#>  [8,]  0.010378412  0.016327303 -0.04663468 -0.18388042 -0.1995020 -0.14426617
#>  [9,] -0.075909150 -0.151379384 -0.03556717 -0.02128133 -0.2713425 -0.02119242
#> [10,]  0.030721163 -0.033675908 -0.04328662  0.07965518 -0.2196458  0.03591447
#>               V41          V42          V43         V44         V45         V46
#>  [1,]  0.02188404  0.082561455 -0.007380457 -0.01495025 -0.06886295 -0.07294033
#>  [2,] -0.13654207  0.003699342 -0.137361475 -0.18193002 -0.23582353 -0.09386458
#>  [3,]  0.00231895  0.012946930 -0.060170489 -0.05797098 -0.11656635 -0.08580426
#>  [4,] -0.04361452 -0.145261166 -0.156697698 -0.04151591 -0.07400851 -0.11943073
#>  [5,] -0.09265971 -0.151671696 -0.165421304 -0.06684594 -0.06729644 -0.21649998
#>  [6,] -0.07890527 -0.058546618 -0.056261501 -0.09984773 -0.13669275 -0.09878115
#>  [7,] -0.12681790 -0.117348290 -0.081174851 -0.14598822 -0.07828113 -0.08448884
#>  [8,] -0.13317910 -0.109673704 -0.115280956 -0.12576872 -0.22748704 -0.22039194
#>  [9,] -0.15488691 -0.071472740 -0.120835588 -0.19024997 -0.28093349 -0.26266748
#> [10,] -0.03341560 -0.036693078  0.032739213 -0.03048471 -0.08379632 -0.16855108
#>              V47        V48        V49         V5        V50        V51
#>  [1,] -0.1156212 -0.1143965 -0.2271232 -0.2807773 -0.3203683 -0.3264805
#>  [2,] -0.2480877 -0.1963348 -0.2091456 -0.2265539 -0.2134281 -0.2191687
#>  [3,] -0.2464749 -0.1848834 -0.2956820 -0.1897496 -0.2997018 -0.4226441
#>  [4,] -0.1188973 -0.2122334 -0.2179059 -0.1528237 -0.2423465 -0.2814536
#>  [5,] -0.2104153 -0.2173754 -0.3071298 -0.2894611 -0.3061551 -0.3990148
#>  [6,] -0.2394095 -0.1708306 -0.2971930 -0.2979956 -0.3217158 -0.3121841
#>  [7,] -0.1195157 -0.2264038 -0.2617695 -0.1082420 -0.3408111 -0.2825212
#>  [8,] -0.1829582 -0.3010664 -0.2288318 -0.2076352 -0.3115411 -0.2255060
#>  [9,] -0.2629247 -0.1607319 -0.2900171 -0.3232101 -0.4201995 -0.4246648
#> [10,] -0.2171167 -0.1967341 -0.1780613 -0.2481657 -0.2524281 -0.2944504
#>              V52        V53        V54        V55        V56        V57
#>  [1,] -0.3865930 -0.4052133 -0.2668438 -0.3796419 -0.3927822 -0.4149732
#>  [2,] -0.2453347 -0.2573428 -0.3766809 -0.2351426 -0.3312233 -0.2149884
#>  [3,] -0.2547978 -0.3064504 -0.3903437 -0.4135153 -0.2704838 -0.2962357
#>  [4,] -0.2327639 -0.3117316 -0.2585055 -0.3137172 -0.3253192 -0.3132891
#>  [5,] -0.3956493 -0.3200990 -0.3040579 -0.2930840 -0.2485928 -0.2688034
#>  [6,] -0.3732543 -0.3214298 -0.2237298 -0.3608989 -0.2996684 -0.3730189
#>  [7,] -0.2188239 -0.3749900 -0.2434646 -0.3004933 -0.3722261 -0.3220504
#>  [8,] -0.3131671 -0.3684064 -0.3942673 -0.2886291 -0.3712523 -0.3846733
#>  [9,] -0.4167074 -0.3701136 -0.3568056 -0.2759006 -0.4644283 -0.3905133
#> [10,] -0.3631999 -0.3406463 -0.3588679 -0.3567617 -0.2091142 -0.2804293
#>              V58        V59         V6        V60         V7          V8
#>  [1,] -0.2721952 -0.3986297 -0.1439068 -0.3118243 -0.2282869 -0.20219602
#>  [2,] -0.3480889 -0.2245406 -0.1516426 -0.3460326 -0.1266552 -0.24240291
#>  [3,] -0.3521381 -0.4076140 -0.2207988 -0.3896213 -0.0698960 -0.08262433
#>  [4,] -0.3827685 -0.3546397 -0.1880101 -0.3592034 -0.2423501 -0.10543818
#>  [5,] -0.2827674 -0.4370174 -0.2530314 -0.3568716 -0.2646712 -0.10639156
#>  [6,] -0.3928977 -0.2849222 -0.1012353 -0.2315785 -0.1847659 -0.19382110
#>  [7,] -0.3267145 -0.2741083 -0.2524752 -0.2983042 -0.0660644 -0.15347260
#>  [8,] -0.2523025 -0.2604371 -0.2510286 -0.2663344 -0.1187351 -0.23385895
#>  [9,] -0.3774511 -0.3181628 -0.2046209 -0.3537623 -0.1566845 -0.27305722
#> [10,] -0.3916121 -0.3873365 -0.2271030 -0.3847341 -0.1768874 -0.16454232
#>                 V9
#>  [1,] -0.043001970
#>  [2,] -0.103627729
#>  [3,] -0.174954457
#>  [4,] -0.015853540
#>  [5,] -0.170017675
#>  [6,] -0.167358814
#>  [7,] -0.147167633
#>  [8,] -0.257801573
#>  [9,] -0.204215140
#> [10,] -0.009316692
#> 
#> $W[[2]]
#>             [,1]         [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,] 0.006870607 -0.088476974 0.01975412 -0.04874288  0.05273352  0.02541553
#> [2,] 0.057314025 -0.006771391 0.08368865 -0.01545981 -0.05258101 -0.01199246
#>             [,7]       [,8]        [,9]       [,10]
#> [1,] -0.02639247 -0.1560371 -0.03029000 -0.09114741
#> [2,] -0.01340726  0.1010035 -0.01126029 -0.05605901
#> 
#> 
#> $vW
#> $vW[[1]]
#>                 V1          V10          V11          V12          V13
#>  [1,] 1.456443e-04 0.0011510628 0.0011764974 0.0012548421 0.0012913525
#>  [2,] 1.210877e-04 0.0009360249 0.0009942676 0.0010867633 0.0011552212
#>  [3,] 1.204634e-04 0.0009450438 0.0009488598 0.0010165207 0.0010356835
#>  [4,] 1.371522e-04 0.0010955793 0.0010965379 0.0012111299 0.0012677438
#>  [5,] 7.882711e-05 0.0007494544 0.0006582808 0.0009342279 0.0010420628
#>  [6,] 5.975776e-05 0.0005424359 0.0004276252 0.0005944415 0.0007651539
#>  [7,] 8.122791e-05 0.0007007075 0.0006378312 0.0007620129 0.0009033667
#>  [8,] 2.265950e-04 0.0016839152 0.0020353076 0.0020666615 0.0019207503
#>  [9,] 1.123715e-04 0.0009211162 0.0009001188 0.0010587470 0.0011289273
#> [10,] 1.064595e-04 0.0008789871 0.0008698727 0.0009483318 0.0009842669
#>               V14          V15          V16          V17          V18
#>  [1,] 0.001310990 0.0015808955 0.0021149374 0.0024279720 0.0029236646
#>  [2,] 0.001174703 0.0013612638 0.0018158214 0.0021145919 0.0024882438
#>  [3,] 0.001060702 0.0012482674 0.0016220488 0.0018175229 0.0021924459
#>  [4,] 0.001329153 0.0016649424 0.0023052962 0.0026453783 0.0031744953
#>  [5,] 0.001214105 0.0019003760 0.0030802908 0.0036028395 0.0043004085
#>  [6,] 0.001040477 0.0017277939 0.0028518720 0.0033582593 0.0039518849
#>  [7,] 0.001111696 0.0016512585 0.0025654113 0.0029958427 0.0035256950
#>  [8,] 0.001480769 0.0008882672 0.0002961762 0.0002632202 0.0003996076
#>  [9,] 0.001222332 0.0016139730 0.0023315252 0.0026877868 0.0032138028
#> [10,] 0.001052605 0.0013367732 0.0018409790 0.0020955114 0.0025432164
#>               V19           V2         V20         V21         V22         V23
#>  [1,] 0.003168310 1.328860e-04 0.003361690 0.003589714 0.003376718 0.003523197
#>  [2,] 0.002689912 1.199385e-04 0.002817057 0.003004803 0.002938119 0.003112767
#>  [3,] 0.002429970 1.132978e-04 0.002619393 0.002799694 0.002602164 0.002676533
#>  [4,] 0.003328434 1.202609e-04 0.003381471 0.003602993 0.003431497 0.003666859
#>  [5,] 0.003903710 3.170835e-05 0.003346238 0.003550885 0.003554873 0.004161960
#>  [6,] 0.003479872 1.835611e-05 0.002937041 0.003111869 0.003120908 0.003745435
#>  [7,] 0.003287742 5.147345e-05 0.002969443 0.003150975 0.003118300 0.003583503
#>  [8,] 0.001721155 2.802483e-04 0.002873994 0.003116234 0.002941912 0.002405631
#>  [9,] 0.003211445 9.026874e-05 0.003093346 0.003267863 0.003116336 0.003399810
#> [10,] 0.002662934 8.985289e-05 0.002731732 0.002891849 0.002706807 0.002866040
#>               V24         V25         V26          V27         V28          V29
#>  [1,] 0.003839099 0.003742501 0.004099150 0.0042552085 0.004345468 0.0042500698
#>  [2,] 0.003386021 0.003341871 0.003582261 0.0036767930 0.003741089 0.0036147499
#>  [3,] 0.002888635 0.002777460 0.003090341 0.0032130629 0.003291789 0.0032465768
#>  [4,] 0.004101160 0.004017598 0.004444257 0.0046264449 0.004684351 0.0046796738
#>  [5,] 0.005159966 0.005164107 0.005942051 0.0063732623 0.006320201 0.0067135891
#>  [6,] 0.004724076 0.004809339 0.005388028 0.0056482151 0.005520325 0.0058158312
#>  [7,] 0.004330882 0.004390513 0.004842480 0.0049968934 0.004885773 0.0050189052
#>  [8,] 0.001587036 0.001274814 0.001056216 0.0009179946 0.001303520 0.0006061949
#>  [9,] 0.003937251 0.003871944 0.004330132 0.0045493313 0.004588486 0.0047147078
#> [10,] 0.003207032 0.003133843 0.003465357 0.0035737560 0.003581882 0.0035922796
#>                 V3          V30           V31           V32           V33
#>  [1,] 1.768771e-04 0.0038088853  0.0034743894  0.0032634016  0.0031348048
#>  [2,] 1.569852e-04 0.0032023443  0.0029388413  0.0027416117  0.0025997118
#>  [3,] 1.487842e-04 0.0029855715  0.0027354893  0.0025502795  0.0024370700
#>  [4,] 1.680503e-04 0.0042300016  0.0039696922  0.0036843684  0.0035419600
#>  [5,] 8.070782e-05 0.0061411364  0.0061639103  0.0056077834  0.0054651148
#>  [6,] 6.460866e-05 0.0052336426  0.0053370729  0.0049087946  0.0047620821
#>  [7,] 9.792895e-05 0.0044886568  0.0044761976  0.0041411950  0.0040007530
#>  [8,] 3.109996e-04 0.0004449469 -0.0004308615 -0.0003372804 -0.0004916751
#>  [9,] 1.354850e-04 0.0042860098  0.0041627984  0.0038453975  0.0037006683
#> [10,] 1.294695e-04 0.0032456921  0.0030707900  0.0028813877  0.0027657728
#>                V34          V35          V36          V37           V38
#>  [1,]  0.003111084  0.003187116  0.003144081  0.002863285  0.0026392968
#>  [2,]  0.002535441  0.002573001  0.002551234  0.002343639  0.0021190979
#>  [3,]  0.002394036  0.002434132  0.002388328  0.002208279  0.0021296772
#>  [4,]  0.003557509  0.003743427  0.003759057  0.003395988  0.0030162753
#>  [5,]  0.005713342  0.006455815  0.006765160  0.005923082  0.0047091062
#>  [6,]  0.004994821  0.005703766  0.005980949  0.005197672  0.0039420559
#>  [7,]  0.004130722  0.004601080  0.004762889  0.004189293  0.0032877092
#>  [8,] -0.001035164 -0.002057363 -0.002523129 -0.001824842 -0.0003696401
#>  [9,]  0.003766675  0.004079146  0.004197315  0.003771368  0.0031989837
#> [10,]  0.002770623  0.002915542  0.002942986  0.002674712  0.0023786950
#>                V39            V4           V40           V41          V42
#>  [1,] 0.0025812353  2.217905e-04  0.0024664032  0.0020747378 0.0017996162
#>  [2,] 0.0020456632  2.031385e-04  0.0019392895  0.0016470398 0.0014416766
#>  [3,] 0.0021100669  1.734139e-04  0.0020143014  0.0017015123 0.0014875950
#>  [4,] 0.0029046923  1.880135e-04  0.0028824642  0.0024014481 0.0020110705
#>  [5,] 0.0042428489 -2.360541e-05  0.0046100447  0.0036693556 0.0027385196
#>  [6,] 0.0034923106 -1.774429e-05  0.0038312444  0.0030556176 0.0022078407
#>  [7,] 0.0029901201  6.331362e-05  0.0031919888  0.0025835373 0.0019474697
#>  [8,] 0.0002144378  5.450692e-04 -0.0005497154 -0.0001673103 0.0005743133
#>  [9,] 0.0030210741  1.173790e-04  0.0031165222  0.0025433001 0.0020366205
#> [10,] 0.0023096900  1.373587e-04  0.0023047756  0.0019023890 0.0015767429
#>               V43         V44          V45          V46          V47
#>  [1,] 0.001559901 0.001408647 0.0011705728 0.0009662592 0.0006816020
#>  [2,] 0.001258396 0.001142143 0.0009574913 0.0007950729 0.0005728563
#>  [3,] 0.001295965 0.001179342 0.0010020093 0.0008157182 0.0005749888
#>  [4,] 0.001676221 0.001514901 0.0011997588 0.0009954380 0.0007116769
#>  [5,] 0.001973100 0.001765061 0.0010764943 0.0009269975 0.0006905293
#>  [6,] 0.001558851 0.001395338 0.0008132714 0.0007218510 0.0005255544
#>  [7,] 0.001459040 0.001315190 0.0008718489 0.0007586173 0.0005481634
#>  [8,] 0.001067388 0.001005993 0.0013344821 0.0010189935 0.0007545010
#>  [9,] 0.001609196 0.001460973 0.0010659044 0.0008922631 0.0006580613
#> [10,] 0.001309285 0.001193339 0.0009304864 0.0007680756 0.0005499212
#>                V48          V49           V5          V50          V51
#>  [1,] 0.0005099281 0.0002634920 0.0003703155 1.057583e-04 8.623450e-05
#>  [2,] 0.0004283584 0.0002234665 0.0003402637 8.741049e-05 7.142178e-05
#>  [3,] 0.0004308400 0.0002259246 0.0002855831 8.894545e-05 6.911974e-05
#>  [4,] 0.0005167770 0.0002566816 0.0003607337 1.053388e-04 8.394652e-05
#>  [5,] 0.0004374012 0.0001641292 0.0002215133 8.743121e-05 6.688577e-05
#>  [6,] 0.0003281986 0.0001127327 0.0002012732 6.629845e-05 4.826795e-05
#>  [7,] 0.0003656390 0.0001516756 0.0002606894 7.436342e-05 5.698267e-05
#>  [8,] 0.0006540083 0.0004304107 0.0005790325 1.304839e-04 1.149964e-04
#>  [9,] 0.0004598748 0.0002131640 0.0002948836 9.087325e-05 7.097814e-05
#> [10,] 0.0003972647 0.0001965941 0.0002668142 8.091213e-05 6.361108e-05
#>                V52          V53          V54          V55          V56
#>  [1,] 7.672627e-05 6.348615e-05 6.589891e-05 4.905598e-05 4.944619e-05
#>  [2,] 6.426292e-05 5.421547e-05 5.488942e-05 4.214006e-05 4.149572e-05
#>  [3,] 5.888563e-05 5.041969e-05 5.319676e-05 3.848240e-05 4.076554e-05
#>  [4,] 7.364878e-05 6.758849e-05 6.787411e-05 5.244942e-05 5.063990e-05
#>  [5,] 5.478747e-05 7.660577e-05 6.966173e-05 6.411310e-05 4.994964e-05
#>  [6,] 4.736233e-05 6.812906e-05 5.934045e-05 5.770944e-05 4.064762e-05
#>  [7,] 5.493341e-05 6.413286e-05 5.851000e-05 5.271279e-05 4.154818e-05
#>  [8,] 9.246962e-05 3.733022e-05 5.330498e-05 1.896238e-05 4.843123e-05
#>  [9,] 5.998565e-05 6.436304e-05 6.285212e-05 5.138136e-05 4.670371e-05
#> [10,] 5.566862e-05 5.204586e-05 5.375643e-05 4.106123e-05 3.999060e-05
#>                V57          V58          V59           V6          V60
#>  [1,] 4.444490e-05 5.494393e-05 4.851902e-05 0.0005342139 4.011863e-05
#>  [2,] 3.780673e-05 4.436159e-05 4.111084e-05 0.0004774773 3.295440e-05
#>  [3,] 3.547994e-05 4.441392e-05 3.943955e-05 0.0004103690 3.274864e-05
#>  [4,] 4.670675e-05 5.690487e-05 5.121982e-05 0.0005545324 4.338151e-05
#>  [5,] 5.340560e-05 6.057026e-05 5.589027e-05 0.0006014660 5.417177e-05
#>  [6,] 4.776430e-05 5.145359e-05 4.770320e-05 0.0005290511 4.480508e-05
#>  [7,] 4.452854e-05 4.982541e-05 4.591696e-05 0.0005218277 4.084297e-05
#>  [8,] 2.373748e-05 3.863989e-05 3.617683e-05 0.0004329877 1.948299e-05
#>  [9,] 4.461703e-05 5.355948e-05 4.883599e-05 0.0005152620 4.232408e-05
#> [10,] 3.684347e-05 4.510616e-05 3.975250e-05 0.0004222863 3.322667e-05
#>                 V7           V8           V9
#>  [1,] 0.0006167368 0.0007470034 0.0010126240
#>  [2,] 0.0005397395 0.0006112889 0.0008097702
#>  [3,] 0.0004925906 0.0006145606 0.0008405554
#>  [4,] 0.0006497452 0.0007656569 0.0009696589
#>  [5,] 0.0007351383 0.0007885352 0.0006872373
#>  [6,] 0.0006176043 0.0006425287 0.0005265913
#>  [7,] 0.0005994019 0.0006417367 0.0006410746
#>  [8,] 0.0004832705 0.0006593797 0.0013545359
#>  [9,] 0.0006210215 0.0007140528 0.0008194345
#> [10,] 0.0005071088 0.0006204708 0.0007823929
#> 
#> $vW[[2]]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,]  0.07896845  0.06550406  0.08668744  0.07485142  0.06601672  0.07143998
#> [2,] -0.07896845 -0.06550406 -0.08668744 -0.07485142 -0.06601672 -0.07143998
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.07514492  0.06499559  0.07904231  0.08779958
#> [2,] -0.07514492 -0.06499559 -0.07904231 -0.08779958
#> 
#> 
#> $B
#> $B[[1]]
#>  [1] 0.5278457 0.2861344 0.6135698 0.3739146 0.5141721 0.2983452 0.4502980
#>  [8] 0.3106734 0.5152589 0.6407484
#> 
#> $B[[2]]
#> [1] -0.058332581 -0.005534519
#> 
#> 
#> $vB
#> $vB[[1]]
#>  [1] 0.0064202874 0.0054002039 0.0049748784 0.0070483032 0.0097185196
#>  [6] 0.0086472875 0.0075865559 0.0009234723 0.0070516071 0.0055430086
#> 
#> $vB[[2]]
#> [1]  0.101097 -0.101097
#> 
#> 
#> $post
#> $post[[1]]
#>           V1    V10    V11    V12    V13    V14    V15    V16    V17    V18
#>  [1,] 0.0272 0.3997 0.3941 0.3309 0.2926 0.1760 0.1739 0.2043 0.2088 0.2678
#>  [2,] 0.0762 0.4459 0.4152 0.3952 0.4256 0.4135 0.4528 0.5326 0.7306 0.6193
#>  [3,] 0.0238 0.2048 0.2652 0.3100 0.2381 0.1918 0.1430 0.1735 0.1781 0.2852
#>  [4,] 0.0067 0.1180 0.0978 0.0909 0.0656 0.0593 0.0832 0.1297 0.2038 0.3811
#>  [5,] 0.0664 0.1838 0.2869 0.4129 0.3647 0.1984 0.2840 0.4039 0.5837 0.6792
#>  [6,] 0.0428 0.0113 0.1255 0.2473 0.3011 0.3747 0.4520 0.5392 0.6588 0.7113
#>  [7,] 0.0139 0.1015 0.1261 0.0828 0.0493 0.0848 0.1514 0.1396 0.1066 0.1923
#>  [8,] 0.0352 0.2912 0.2328 0.2237 0.2470 0.1560 0.3491 0.3308 0.2299 0.2203
#>  [9,] 0.0315 0.1533 0.2128 0.2536 0.2686 0.2803 0.1886 0.1485 0.2160 0.2417
#> [10,] 0.1021 0.1369 0.2509 0.2631 0.2796 0.2977 0.3823 0.3129 0.3956 0.2093
#> [11,] 0.0123 0.0475 0.1152 0.0520 0.1192 0.1943 0.1840 0.2077 0.1956 0.1630
#> [12,] 0.0308 0.2795 0.3336 0.2940 0.1608 0.3335 0.4985 0.7295 0.7350 0.8253
#> [13,] 0.0340 0.4117 0.5245 0.5341 0.5554 0.3915 0.2950 0.3075 0.3021 0.2719
#> [14,] 0.0368 0.2544 0.2936 0.2935 0.2657 0.3187 0.2794 0.2534 0.1980 0.1929
#> [15,] 0.0093 0.1360 0.2132 0.3738 0.3738 0.2673 0.2333 0.5367 0.7312 0.7659
#> [16,] 0.0162 0.1779 0.2164 0.2568 0.3089 0.3829 0.4393 0.5335 0.5996 0.6728
#> [17,] 0.0086 0.1185 0.0775 0.1101 0.1042 0.0853 0.0456 0.1304 0.2690 0.2947
#> [18,] 0.0094 0.1725 0.2228 0.3106 0.4144 0.5157 0.5369 0.5107 0.6441 0.7326
#> [19,] 0.0231 0.2461 0.2245 0.1520 0.1732 0.3099 0.4380 0.5595 0.6820 0.6164
#> [20,] 0.0202 0.1370 0.0843 0.0269 0.1254 0.3046 0.5584 0.7973 0.8341 0.8057
#> [21,] 0.0721 0.0795 0.2534 0.3920 0.3375 0.1610 0.1889 0.3308 0.2282 0.2177
#> [22,] 0.0265 0.0779 0.0327 0.2060 0.1908 0.1065 0.1457 0.2232 0.2070 0.1105
#> [23,] 0.0217 0.2110 0.2343 0.2087 0.1645 0.1689 0.1650 0.1967 0.2934 0.3709
#> [24,] 0.0036 0.0888 0.0937 0.1245 0.1599 0.1542 0.1846 0.1732 0.1477 0.1748
#> [25,] 0.0107 0.2936 0.3104 0.3431 0.2456 0.1887 0.1184 0.2080 0.2736 0.3274
#> [26,] 0.0303 0.2354 0.2898 0.2812 0.1578 0.0273 0.0673 0.1444 0.2070 0.2645
#> [27,] 0.1313 0.2231 0.2907 0.2259 0.3136 0.3302 0.3660 0.3956 0.4386 0.4670
#> [28,] 0.0079 0.1240 0.1097 0.1215 0.1874 0.3383 0.3227 0.2723 0.3943 0.6432
#> [29,] 0.0443 0.3083 0.3794 0.5364 0.6173 0.7842 0.8392 0.9016 1.0000 0.8911
#> [30,] 0.0257 0.0561 0.0891 0.0861 0.1531 0.1524 0.1849 0.2871 0.2009 0.2748
#> [31,] 0.0116 0.2271 0.3171 0.2882 0.2657 0.2307 0.1889 0.1791 0.2298 0.3715
#> [32,] 0.0201 0.2571 0.2931 0.3108 0.3603 0.3002 0.2718 0.2007 0.1801 0.2234
#> [33,] 0.0197 0.1695 0.1734 0.2470 0.3141 0.3297 0.2759 0.2056 0.1162 0.1884
#> [34,] 0.0311 0.0767 0.1767 0.2555 0.2812 0.2722 0.3227 0.3463 0.5395 0.7911
#> [35,] 0.0115 0.0734 0.0740 0.0622 0.1055 0.1183 0.1721 0.2584 0.3232 0.3817
#> [36,] 0.0065 0.1109 0.0937 0.0827 0.0920 0.0911 0.1487 0.1666 0.1268 0.1374
#> [37,] 0.0084 0.1753 0.1970 0.1167 0.1683 0.0814 0.2179 0.5121 0.7231 0.7776
#> [38,] 0.0635 0.3900 0.4674 0.5632 0.5506 0.4343 0.3052 0.3492 0.3975 0.3875
#> [39,] 0.0100 0.1264 0.0881 0.1992 0.0184 0.2261 0.1729 0.2131 0.0693 0.2281
#> [40,] 0.0216 0.0866 0.1564 0.0780 0.0997 0.0915 0.0662 0.1134 0.1740 0.2573
#>          V19     V2    V20    V21    V22    V23    V24    V25    V26    V27
#>  [1,] 0.2434 0.0378 0.1839 0.2802 0.6172 0.8015 0.8313 0.8440 0.8494 0.9168
#>  [2,] 0.2032 0.0666 0.4636 0.4148 0.4292 0.5730 0.5399 0.3161 0.2285 0.6995
#>  [3,] 0.5036 0.0318 0.6166 0.7616 0.8125 0.7793 0.8788 0.8813 0.9470 1.0000
#>  [4,] 0.4451 0.0096 0.5224 0.5911 0.6566 0.6308 0.5998 0.4958 0.5647 0.6906
#>  [5,] 0.6086 0.0575 0.4858 0.3246 0.2013 0.2082 0.1686 0.2484 0.2736 0.2984
#>  [6,] 0.7602 0.0555 0.8672 0.8416 0.7974 0.8385 0.9317 0.8555 0.6162 0.4139
#>  [7,] 0.2991 0.0222 0.3247 0.3797 0.5658 0.7483 0.8757 0.9048 0.7511 0.6858
#>  [8,] 0.2493 0.0116 0.4128 0.3158 0.6191 0.5854 0.3395 0.2561 0.5599 0.8145
#>  [9,] 0.2989 0.0252 0.3341 0.3786 0.3956 0.5232 0.6913 0.7868 0.8337 0.9199
#> [10,] 0.3218 0.0830 0.3345 0.3184 0.2887 0.3610 0.2566 0.4106 0.4591 0.4722
#> [11,] 0.1218 0.0022 0.1017 0.1354 0.3157 0.4645 0.5906 0.6776 0.8119 0.8594
#> [12,] 0.8793 0.0339 0.9657 1.0000 0.8707 0.6471 0.5973 0.8218 0.7755 0.6111
#> [13,] 0.5443 0.0625 0.7932 0.8751 0.8667 0.7107 0.6911 0.7287 0.8792 1.0000
#> [14,] 0.2826 0.0279 0.3245 0.3504 0.3324 0.4217 0.4774 0.4808 0.6325 0.8334
#> [15,] 0.6271 0.0269 0.4395 0.4330 0.4326 0.5544 0.7360 0.8589 0.8989 0.9420
#> [16,] 0.7309 0.0041 0.8092 0.8941 0.9668 1.0000 0.9893 0.9376 0.8991 0.9184
#> [17,] 0.3669 0.0215 0.4948 0.6275 0.8162 0.9237 0.8710 0.8052 0.8756 1.0000
#> [18,] 0.8164 0.0333 0.8856 0.9891 1.0000 0.8750 0.8631 0.9074 0.8674 0.7750
#> [19,] 0.6803 0.0315 0.8435 0.9921 1.0000 0.7983 0.5426 0.3952 0.5179 0.5650
#> [20,] 0.8616 0.0104 0.8769 0.9413 0.9403 0.9409 1.0000 0.9725 0.9309 0.9351
#> [21,] 0.1853 0.1574 0.5167 0.5342 0.6298 0.8437 0.6756 0.5825 0.6141 0.8809
#> [22,] 0.1078 0.0440 0.1165 0.2224 0.0689 0.2060 0.2384 0.0904 0.2278 0.5872
#> [23,] 0.4309 0.0152 0.4161 0.5116 0.6501 0.7717 0.8491 0.9104 0.8912 0.8189
#> [24,] 0.1455 0.0078 0.1579 0.2257 0.1975 0.3368 0.5828 0.8505 1.0000 0.8457
#> [25,] 0.2344 0.0453 0.1260 0.0576 0.1241 0.3239 0.4357 0.5734 0.7825 0.9252
#> [26,] 0.2828 0.0353 0.4293 0.5685 0.6990 0.7246 0.7622 0.9242 1.0000 0.9979
#> [27,] 0.5255 0.2339 0.3735 0.2243 0.1973 0.4337 0.6532 0.5070 0.2796 0.4163
#> [28,] 0.7271 0.0086 0.8673 0.9674 0.9847 0.9480 0.8036 0.6833 0.5136 0.3090
#> [29,] 0.8753 0.0446 0.7886 0.7156 0.7581 0.6372 0.3210 0.2076 0.2279 0.3309
#> [30,] 0.5017 0.0447 0.2172 0.4978 0.5265 0.3647 0.5768 0.5161 0.5715 0.4006
#> [31,] 0.6223 0.0744 0.7260 0.7934 0.8045 0.8067 0.9173 0.9327 0.9562 1.0000
#> [32,] 0.3568 0.0165 0.5492 0.7209 0.8318 0.8864 0.9520 0.9637 1.0000 0.9673
#> [33,] 0.3390 0.0394 0.3926 0.4282 0.5418 0.6448 0.7223 0.7853 0.7984 0.8847
#> [34,] 0.9064 0.0491 0.8701 0.7672 0.2957 0.4148 0.6043 0.3178 0.3482 0.6158
#> [35,] 0.4243 0.0150 0.4217 0.4449 0.4075 0.3306 0.4012 0.4466 0.5218 0.7552
#> [36,] 0.1095 0.0122 0.1286 0.2146 0.2889 0.4238 0.6168 0.8167 0.9622 0.8280
#> [37,] 0.6222 0.0153 0.3501 0.3733 0.2622 0.3776 0.7361 0.8673 0.8223 0.7772
#> [38,] 0.5280 0.0709 0.7198 0.7702 0.8562 0.8688 0.9236 1.0000 0.9662 0.9822
#> [39,] 0.4060 0.0171 0.3973 0.2741 0.3690 0.5556 0.4846 0.3140 0.5334 0.5256
#> [40,] 0.3294 0.0124 0.3910 0.5438 0.6115 0.7022 0.7610 0.7973 0.9105 0.8807
#>          V28    V29     V3    V30    V31    V32    V33    V34    V35    V36
#>  [1,] 1.0000 0.7896 0.0488 0.5371 0.6472 0.6505 0.4959 0.2175 0.0990 0.0434
#>  [2,] 1.0000 0.7262 0.0481 0.4724 0.5103 0.5459 0.2881 0.0981 0.1951 0.4181
#>  [3,] 0.9739 0.8446 0.0422 0.6151 0.4302 0.3165 0.2869 0.2017 0.1206 0.0271
#>  [4,] 0.8513 1.0000 0.0024 0.9166 0.7676 0.6177 0.5468 0.5516 0.5463 0.5515
#>  [5,] 0.4655 0.6990 0.0842 0.7474 0.7956 0.7981 0.6715 0.6942 0.7440 0.8169
#>  [6,] 0.3269 0.3108 0.0708 0.2554 0.3367 0.4465 0.5000 0.5111 0.5194 0.4619
#>  [7,] 0.7043 0.5864 0.0089 0.3773 0.2206 0.2628 0.2672 0.2907 0.1982 0.2288
#>  [8,] 0.6941 0.6985 0.0191 0.8660 0.5930 0.3664 0.6750 0.8697 0.7837 0.7552
#>  [9,] 1.0000 0.8990 0.0167 0.6456 0.5967 0.4355 0.2997 0.2294 0.1866 0.0922
#> [10,] 0.7278 0.7591 0.0577 0.6579 0.7514 0.6666 0.4903 0.5962 0.6552 0.4014
#> [11,] 0.9228 0.8387 0.0196 0.7238 0.6292 0.5181 0.4629 0.5255 0.5147 0.3929
#> [12,] 0.4195 0.2990 0.0202 0.1354 0.2438 0.5624 0.5555 0.6963 0.7298 0.7022
#> [13,] 0.9816 0.8984 0.0381 0.6048 0.4934 0.5371 0.4586 0.2908 0.0774 0.2249
#> [14,] 0.9458 1.0000 0.0103 0.8425 0.5524 0.4795 0.5200 0.3968 0.1940 0.1519
#> [15,] 0.9401 0.9379 0.0217 0.8575 0.7284 0.6700 0.7547 0.8773 0.9919 0.9922
#> [16,] 0.9128 0.7811 0.0239 0.6018 0.3765 0.3300 0.2280 0.0212 0.1117 0.1788
#> [17,] 0.9858 0.9427 0.0242 0.8114 0.6987 0.6810 0.6591 0.6954 0.7290 0.6680
#> [18,] 0.6600 0.5615 0.0306 0.4016 0.2331 0.1164 0.1095 0.0431 0.0619 0.1956
#> [19,] 0.3042 0.1881 0.0170 0.3960 0.2286 0.3544 0.4187 0.2398 0.1847 0.3760
#> [20,] 0.7317 0.4421 0.0325 0.3244 0.4161 0.4611 0.4031 0.3000 0.2459 0.1348
#> [21,] 0.8375 0.3869 0.1112 0.5051 0.5455 0.4241 0.1534 0.4950 0.6983 0.7109
#> [22,] 0.8457 0.8467 0.0137 0.7679 0.8055 0.6260 0.6545 0.8747 0.9885 0.9348
#> [23,] 0.6779 0.5368 0.0346 0.5207 0.5651 0.5749 0.5250 0.4255 0.3330 0.2331
#> [24,] 0.6624 0.5564 0.0092 0.3925 0.3233 0.2054 0.1920 0.2227 0.3147 0.2268
#> [25,] 0.9349 0.9348 0.0289 1.0000 0.9308 0.8478 0.7605 0.7040 0.7539 0.7990
#> [26,] 0.8297 0.7032 0.0490 0.7141 0.6893 0.4961 0.2584 0.0969 0.0776 0.0364
#> [27,] 0.5950 0.5242 0.3059 0.4178 0.3714 0.2375 0.0863 0.1437 0.2896 0.4577
#> [28,] 0.0832 0.4019 0.0055 0.2344 0.1905 0.1235 0.1717 0.2351 0.2489 0.3649
#> [29,] 0.2847 0.1949 0.0235 0.1671 0.1025 0.1362 0.2212 0.1124 0.1677 0.1039
#> [30,] 0.3650 0.6685 0.0388 0.8659 0.8052 0.4082 0.3379 0.5092 0.6776 0.7313
#> [31,] 0.9818 0.8684 0.0367 0.6381 0.3997 0.3242 0.2835 0.2413 0.2321 0.1260
#> [32,] 0.8664 0.7896 0.0344 0.6345 0.5351 0.4056 0.2563 0.2894 0.3588 0.4296
#> [33,] 0.9582 0.8990 0.0384 0.6831 0.6108 0.5480 0.5058 0.4476 0.2401 0.1405
#> [34,] 0.8049 0.6289 0.0692 0.4999 0.5830 0.6660 0.4124 0.1260 0.2487 0.4676
#> [35,] 0.9503 1.0000 0.0136 0.9084 0.8283 0.7571 0.7262 0.6152 0.5680 0.5757
#> [36,] 0.5816 0.4667 0.0068 0.3539 0.2727 0.1410 0.1863 0.2176 0.2360 0.1725
#> [37,] 0.7862 0.5652 0.0291 0.3635 0.3534 0.3865 0.3370 0.1693 0.2627 0.3195
#> [38,] 0.7360 0.4158 0.0453 0.2918 0.3280 0.3690 0.3450 0.2863 0.0864 0.3724
#> [39,] 0.2520 0.2090 0.0623 0.3559 0.6260 0.7340 0.6120 0.3497 0.3953 0.3012
#> [40,] 0.7949 0.7990 0.0174 0.7180 0.6407 0.6312 0.5929 0.6168 0.6498 0.6764
#>          V37    V38    V39     V4    V40    V41    V42    V43    V44    V45
#>  [1,] 0.1708 0.1979 0.1880 0.0848 0.1108 0.1702 0.0585 0.0638 0.1391 0.0638
#>  [2,] 0.4604 0.3217 0.2828 0.0394 0.2430 0.1979 0.2444 0.1847 0.0841 0.0692
#>  [3,] 0.0580 0.1262 0.1072 0.0399 0.1082 0.0360 0.1197 0.2061 0.2054 0.1878
#>  [4,] 0.4561 0.3466 0.3384 0.0058 0.2853 0.2502 0.1641 0.1605 0.1491 0.1326
#>  [5,] 0.8912 1.0000 0.8753 0.0372 0.7061 0.6803 0.5898 0.4618 0.3639 0.1492
#>  [6,] 0.4234 0.4372 0.4277 0.0618 0.4433 0.3700 0.3324 0.2564 0.2527 0.2137
#>  [7,] 0.3186 0.2871 0.2921 0.0108 0.2806 0.2682 0.2112 0.1513 0.1789 0.1850
#>  [8,] 0.5789 0.4713 0.1252 0.0469 0.6087 0.7322 0.5977 0.3431 0.1803 0.2378
#>  [9,] 0.1829 0.1743 0.2452 0.0479 0.2407 0.2518 0.3184 0.1685 0.0675 0.1186
#> [10,] 0.1188 0.3245 0.3107 0.0627 0.1354 0.5109 0.7988 0.7517 0.5508 0.5858
#> [11,] 0.1279 0.0411 0.0859 0.0206 0.1131 0.1306 0.1757 0.2648 0.1955 0.0656
#> [12,] 0.5468 0.1421 0.4738 0.0889 0.6410 0.4375 0.3178 0.2377 0.2808 0.1374
#> [13,] 0.1602 0.3958 0.6117 0.0257 0.5196 0.2321 0.4370 0.3797 0.4322 0.4892
#> [14,] 0.2010 0.1736 0.1029 0.0566 0.2244 0.3717 0.4449 0.3939 0.2030 0.2010
#> [15,] 0.9419 0.8388 0.6605 0.0339 0.4816 0.2917 0.1769 0.1136 0.0701 0.1578
#> [16,] 0.2373 0.2843 0.2241 0.0441 0.2715 0.3363 0.2546 0.1867 0.2160 0.1278
#> [17,] 0.5917 0.4899 0.3439 0.0445 0.2366 0.1716 0.1013 0.0766 0.0845 0.0260
#> [18,] 0.2120 0.3242 0.4102 0.0376 0.2939 0.1911 0.1702 0.1010 0.1512 0.1427
#> [19,] 0.4331 0.3626 0.2519 0.0226 0.1870 0.1046 0.2339 0.1991 0.1100 0.0684
#> [20,] 0.2541 0.2255 0.1598 0.0239 0.1485 0.0845 0.0569 0.0855 0.1262 0.1153
#> [21,] 0.5647 0.4870 0.5515 0.1085 0.4433 0.5250 0.6075 0.5251 0.1359 0.4268
#> [22,] 0.6960 0.5733 0.5872 0.0084 0.6663 0.5651 0.5247 0.3684 0.1997 0.1512
#> [23,] 0.1451 0.1648 0.2694 0.0346 0.3730 0.4467 0.4133 0.3743 0.3021 0.2069
#> [24,] 0.0795 0.0748 0.1166 0.0387 0.1969 0.2619 0.2507 0.1983 0.0948 0.0931
#> [25,] 0.7673 0.5955 0.4731 0.0713 0.4840 0.4340 0.3954 0.4837 0.5379 0.4485
#> [26,] 0.1572 0.1823 0.1349 0.0608 0.0849 0.0492 0.1367 0.1552 0.1548 0.1319
#> [27,] 0.3725 0.3372 0.3803 0.4264 0.4181 0.3603 0.2711 0.1653 0.1951 0.2811
#> [28,] 0.3382 0.1589 0.0989 0.0250 0.1089 0.1043 0.0839 0.1391 0.0819 0.0678
#> [29,] 0.2562 0.2624 0.2236 0.1008 0.1180 0.1103 0.2831 0.2385 0.0255 0.1967
#> [30,] 0.6062 0.7040 0.8849 0.0239 0.8979 0.7751 0.7247 0.7733 0.7762 0.6009
#> [31,] 0.0693 0.0701 0.1439 0.0225 0.1475 0.0438 0.0469 0.1476 0.1742 0.1555
#> [32,] 0.4773 0.4516 0.3765 0.0330 0.3051 0.1921 0.1184 0.1984 0.1570 0.0660
#> [33,] 0.1772 0.1742 0.3326 0.0076 0.4021 0.3009 0.2075 0.1206 0.0255 0.0298
#> [34,] 0.5382 0.3150 0.2139 0.0831 0.1848 0.1679 0.2328 0.1015 0.0713 0.0615
#> [35,] 0.5324 0.3672 0.1669 0.0076 0.0866 0.0646 0.1891 0.2683 0.2887 0.2341
#> [36,] 0.0589 0.0621 0.1847 0.0108 0.2452 0.2984 0.3041 0.2275 0.1480 0.1102
#> [37,] 0.1388 0.1048 0.1681 0.0432 0.1910 0.1174 0.0933 0.0856 0.0951 0.0986
#> [38,] 0.4649 0.3488 0.1817 0.0333 0.1142 0.1220 0.2621 0.4461 0.4726 0.3263
#> [39,] 0.5408 0.8814 0.9857 0.0205 0.9167 0.6121 0.5006 0.3210 0.3202 0.4295
#> [40,] 0.6253 0.5117 0.3890 0.0152 0.3273 0.2509 0.1530 0.1323 0.1657 0.1215
#>          V46    V47    V48    V49     V5    V50    V51    V52    V53    V54
#>  [1,] 0.0581 0.0641 0.1044 0.0732 0.1127 0.0275 0.0146 0.0091 0.0045 0.0043
#>  [2,] 0.0528 0.0357 0.0085 0.0230 0.0590 0.0046 0.0156 0.0031 0.0054 0.0105
#>  [3,] 0.2047 0.1716 0.1069 0.0477 0.0788 0.0170 0.0186 0.0096 0.0071 0.0084
#>  [4,] 0.0687 0.0602 0.0561 0.0306 0.0197 0.0154 0.0029 0.0048 0.0023 0.0020
#>  [5,] 0.1216 0.1306 0.1198 0.0578 0.0458 0.0235 0.0135 0.0141 0.0190 0.0043
#>  [6,] 0.1789 0.1010 0.0528 0.0453 0.1215 0.0118 0.0009 0.0142 0.0179 0.0079
#>  [7,] 0.1717 0.0898 0.0656 0.0445 0.0215 0.0110 0.0024 0.0062 0.0072 0.0113
#>  [8,] 0.3424 0.2303 0.0689 0.0216 0.0737 0.0469 0.0426 0.0346 0.0158 0.0154
#>  [9,] 0.1833 0.1878 0.1114 0.0310 0.0902 0.0143 0.0138 0.0108 0.0062 0.0044
#> [10,] 0.7292 0.5522 0.3339 0.1608 0.0635 0.0475 0.1004 0.0709 0.0317 0.0309
#> [11,] 0.0580 0.0319 0.0301 0.0272 0.0180 0.0074 0.0149 0.0125 0.0134 0.0026
#> [12,] 0.1136 0.1034 0.0688 0.0422 0.1570 0.0117 0.0070 0.0167 0.0127 0.0138
#> [13,] 0.1901 0.0940 0.1364 0.0906 0.0441 0.0144 0.0329 0.0141 0.0019 0.0067
#> [14,] 0.2187 0.1840 0.1477 0.0971 0.0759 0.0224 0.0151 0.0105 0.0024 0.0018
#> [15,] 0.1938 0.1106 0.0693 0.0176 0.0305 0.0205 0.0309 0.0212 0.0091 0.0056
#> [16,] 0.0768 0.1070 0.0946 0.0636 0.0630 0.0227 0.0128 0.0173 0.0135 0.0114
#> [17,] 0.0333 0.0205 0.0309 0.0101 0.0667 0.0095 0.0047 0.0072 0.0054 0.0022
#> [18,] 0.1097 0.1173 0.0972 0.0703 0.1296 0.0281 0.0216 0.0153 0.0112 0.0241
#> [19,] 0.0303 0.0674 0.0785 0.0455 0.0410 0.0246 0.0151 0.0125 0.0036 0.0123
#> [20,] 0.0570 0.0426 0.0425 0.0235 0.0807 0.0006 0.0188 0.0127 0.0081 0.0067
#> [21,] 0.4442 0.2193 0.0900 0.1200 0.0666 0.0628 0.0234 0.0309 0.0127 0.0082
#> [22,] 0.0508 0.0931 0.0982 0.0524 0.0305 0.0188 0.0100 0.0038 0.0187 0.0156
#> [23,] 0.1790 0.1689 0.1341 0.0769 0.0484 0.0222 0.0205 0.0123 0.0067 0.0011
#> [24,] 0.0965 0.0381 0.0435 0.0336 0.0530 0.0055 0.0079 0.0119 0.0055 0.0035
#> [25,] 0.2674 0.1541 0.1359 0.0941 0.1075 0.0261 0.0079 0.0164 0.0120 0.0113
#> [26,] 0.0985 0.1258 0.0954 0.0489 0.0167 0.0241 0.0042 0.0086 0.0046 0.0126
#> [27,] 0.2246 0.1921 0.1500 0.0665 0.4010 0.0193 0.0156 0.0362 0.0210 0.0154
#> [28,] 0.0663 0.1202 0.0692 0.0152 0.0344 0.0266 0.0174 0.0176 0.0127 0.0088
#> [29,] 0.1483 0.0434 0.0627 0.0513 0.2252 0.0473 0.0248 0.0274 0.0205 0.0141
#> [30,] 0.4514 0.3096 0.1859 0.0956 0.1315 0.0206 0.0206 0.0096 0.0153 0.0096
#> [31,] 0.1651 0.1181 0.0720 0.0321 0.0076 0.0056 0.0202 0.0141 0.0103 0.0100
#> [32,] 0.1294 0.0797 0.0052 0.0233 0.0397 0.0152 0.0125 0.0054 0.0057 0.0137
#> [33,] 0.0691 0.0781 0.0777 0.0369 0.0251 0.0057 0.0091 0.0134 0.0097 0.0042
#> [34,] 0.0779 0.0761 0.0845 0.0592 0.0079 0.0068 0.0089 0.0087 0.0032 0.0130
#> [35,] 0.1668 0.1015 0.1195 0.0704 0.0211 0.0167 0.0107 0.0091 0.0016 0.0084
#> [36,] 0.1178 0.0608 0.0333 0.0276 0.0217 0.0100 0.0023 0.0069 0.0025 0.0027
#> [37,] 0.0956 0.0426 0.0407 0.0106 0.0951 0.0179 0.0056 0.0236 0.0114 0.0136
#> [38,] 0.1423 0.0390 0.0406 0.0311 0.0185 0.0086 0.0154 0.0048 0.0025 0.0087
#> [39,] 0.3654 0.2655 0.1576 0.0681 0.0205 0.0294 0.0241 0.0121 0.0036 0.0150
#> [40,] 0.0978 0.0452 0.0273 0.0179 0.0608 0.0092 0.0018 0.0052 0.0049 0.0096
#>          V55    V56    V57    V58    V59     V6    V60     V7     V8     V9
#>  [1,] 0.0043 0.0098 0.0054 0.0051 0.0065 0.1103 0.0103 0.1349 0.2337 0.3113
#>  [2,] 0.0110 0.0015 0.0072 0.0048 0.0107 0.0649 0.0094 0.1209 0.2467 0.3564
#>  [3,] 0.0038 0.0026 0.0028 0.0013 0.0035 0.0766 0.0060 0.0881 0.1143 0.1594
#>  [4,] 0.0040 0.0019 0.0034 0.0034 0.0051 0.0618 0.0031 0.0432 0.0951 0.0836
#>  [5,] 0.0036 0.0026 0.0024 0.0162 0.0109 0.0771 0.0079 0.0771 0.1130 0.2353
#>  [6,] 0.0060 0.0131 0.0089 0.0084 0.0113 0.1524 0.0049 0.1543 0.0391 0.0610
#>  [7,] 0.0012 0.0022 0.0025 0.0059 0.0039 0.0136 0.0048 0.0659 0.0954 0.0786
#>  [8,] 0.0109 0.0048 0.0095 0.0015 0.0073 0.1185 0.0067 0.1683 0.1541 0.1466
#>  [9,] 0.0072 0.0007 0.0054 0.0035 0.0001 0.1057 0.0055 0.1024 0.1209 0.1241
#> [10,] 0.0252 0.0087 0.0177 0.0214 0.0227 0.1328 0.0106 0.0988 0.1787 0.1199
#> [11,] 0.0038 0.0018 0.0113 0.0058 0.0047 0.0492 0.0071 0.0033 0.0398 0.0791
#> [12,] 0.0090 0.0051 0.0029 0.0122 0.0056 0.1750 0.0020 0.0920 0.1353 0.1593
#> [13,] 0.0099 0.0042 0.0057 0.0051 0.0033 0.1027 0.0058 0.1287 0.1850 0.2647
#> [14,] 0.0057 0.0092 0.0009 0.0086 0.0110 0.0679 0.0052 0.0970 0.1473 0.2164
#> [15,] 0.0086 0.0092 0.0070 0.0116 0.0060 0.1172 0.0110 0.1450 0.0638 0.0740
#> [16,] 0.0062 0.0157 0.0088 0.0036 0.0053 0.0921 0.0030 0.1368 0.1078 0.1552
#> [17,] 0.0016 0.0029 0.0058 0.0050 0.0024 0.0771 0.0030 0.0499 0.0906 0.1229
#> [18,] 0.0164 0.0055 0.0078 0.0055 0.0091 0.1795 0.0067 0.1909 0.1692 0.1870
#> [19,] 0.0043 0.0114 0.0052 0.0091 0.0008 0.0116 0.0092 0.0223 0.0805 0.2365
#> [20,] 0.0043 0.0065 0.0049 0.0054 0.0073 0.1529 0.0054 0.1154 0.0608 0.1317
#> [21,] 0.0281 0.0117 0.0092 0.0147 0.0157 0.1800 0.0129 0.1108 0.2794 0.1408
#> [22,] 0.0068 0.0097 0.0073 0.0081 0.0086 0.0438 0.0095 0.0341 0.0780 0.0844
#> [23,] 0.0026 0.0049 0.0029 0.0022 0.0022 0.0526 0.0032 0.0773 0.0862 0.1451
#> [24,] 0.0036 0.0004 0.0018 0.0049 0.0024 0.1197 0.0016 0.1243 0.1026 0.1239
#> [25,] 0.0021 0.0097 0.0072 0.0060 0.0017 0.1019 0.0036 0.1606 0.2119 0.3061
#> [26,] 0.0036 0.0035 0.0034 0.0079 0.0036 0.1354 0.0048 0.1465 0.1123 0.1945
#> [27,] 0.0180 0.0013 0.0106 0.0127 0.0178 0.1791 0.0231 0.1853 0.0055 0.1929
#> [28,] 0.0098 0.0019 0.0059 0.0058 0.0059 0.0546 0.0032 0.0528 0.0958 0.1009
#> [29,] 0.0185 0.0055 0.0045 0.0115 0.0152 0.2611 0.0100 0.2061 0.1668 0.1801
#> [30,] 0.0131 0.0198 0.0025 0.0199 0.0255 0.1323 0.0180 0.1608 0.2145 0.0847
#> [31,] 0.0034 0.0026 0.0037 0.0044 0.0057 0.0545 0.0035 0.1110 0.1069 0.1708
#> [32,] 0.0109 0.0035 0.0056 0.0105 0.0082 0.0443 0.0036 0.0684 0.0903 0.1739
#> [33,] 0.0058 0.0072 0.0041 0.0045 0.0047 0.0629 0.0054 0.0747 0.0578 0.1357
#> [34,] 0.0188 0.0101 0.0229 0.0182 0.0046 0.0200 0.0038 0.0981 0.1016 0.2025
#> [35,] 0.0064 0.0026 0.0029 0.0037 0.0070 0.1058 0.0041 0.1023 0.0440 0.0931
#> [36,] 0.0052 0.0036 0.0026 0.0036 0.0006 0.0284 0.0035 0.0527 0.0575 0.1054
#> [37,] 0.0117 0.0060 0.0058 0.0031 0.0072 0.0752 0.0045 0.0414 0.0259 0.0692
#> [38,] 0.0072 0.0095 0.0086 0.0085 0.0040 0.1260 0.0051 0.1015 0.1918 0.3362
#> [39,] 0.0085 0.0073 0.0050 0.0044 0.0040 0.0368 0.0117 0.1098 0.1276 0.0598
#> [40,] 0.0134 0.0122 0.0047 0.0018 0.0006 0.1026 0.0023 0.1139 0.0877 0.1160
#> 
#> $post[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.7772574 0.6063694 0.8605509 0.7714687 0.7919492 0.6537024 0.6638238
#>  [2,] 0.7383504 0.5258588 0.8266932 0.6874662 0.6869240 0.5933325 0.6220284
#>  [3,] 0.8168751 0.6330802 0.8949601 0.7953283 0.8487525 0.6983756 0.6729445
#>  [4,] 0.8090088 0.6074887 0.8774609 0.7757128 0.7635050 0.6833142 0.7127194
#>  [5,] 0.7106043 0.3898823 0.7658101 0.5885255 0.4267780 0.5000485 0.5639341
#>  [6,] 0.7684884 0.4831586 0.8574515 0.7076102 0.7470930 0.6134481 0.5670839
#>  [7,] 0.7922979 0.6123679 0.8660559 0.7573525 0.7979441 0.6720987 0.6590902
#>  [8,] 0.7485486 0.4901312 0.8224782 0.6605055 0.5672614 0.5706444 0.5799971
#>  [9,] 0.7948489 0.6192188 0.8657678 0.7524916 0.7724744 0.6577266 0.6675726
#> [10,] 0.6691370 0.3591894 0.7099753 0.5457307 0.4411344 0.4335745 0.4923755
#> [11,] 0.7891886 0.6476811 0.8637825 0.7651821 0.7588894 0.6773960 0.6949524
#> [12,] 0.7586824 0.4616618 0.8685511 0.7221960 0.7302751 0.6103308 0.5365430
#> [13,] 0.8112796 0.5175504 0.8799732 0.7355827 0.7668889 0.6219826 0.6045965
#> [14,] 0.7622240 0.5642941 0.8282981 0.7033279 0.6899579 0.6073738 0.6262348
#> [15,] 0.8000659 0.5554965 0.9054382 0.7703686 0.6818151 0.6602163 0.6739593
#> [16,] 0.8276498 0.6112942 0.9109066 0.7962363 0.8632048 0.7071869 0.6497100
#> [17,] 0.8392506 0.6436991 0.9159614 0.8241697 0.8174065 0.7238661 0.7354503
#> [18,] 0.7964934 0.5573928 0.8896418 0.7582849 0.8335353 0.6652866 0.5951068
#> [19,] 0.7683872 0.5345683 0.8475547 0.7241837 0.7830584 0.6540636 0.6046538
#> [20,] 0.8234250 0.6240973 0.9081060 0.8082922 0.8783184 0.7344782 0.6606986
#> [21,] 0.7559020 0.4388026 0.8318224 0.6471763 0.6003245 0.5366271 0.5672456
#> [22,] 0.7454725 0.4974500 0.7948955 0.6365103 0.4790367 0.5530443 0.6403996
#> [23,] 0.7987297 0.5684421 0.8649897 0.7437377 0.7769159 0.6486043 0.6237057
#> [24,] 0.7621810 0.6372977 0.8528287 0.7287460 0.7598264 0.6606207 0.6407283
#> [25,] 0.7450413 0.4829363 0.8287272 0.6787789 0.5339012 0.5563393 0.6234162
#> [26,] 0.8050125 0.6456646 0.8811480 0.7808886 0.8325242 0.6966437 0.6885209
#> [27,] 0.6299169 0.4039305 0.7012846 0.5504008 0.5055605 0.4451106 0.4988654
#> [28,] 0.7634252 0.5557889 0.8581458 0.7431104 0.8216140 0.6674887 0.6004099
#> [29,] 0.6556954 0.4099100 0.7629656 0.6116742 0.6452358 0.5287384 0.4737998
#> [30,] 0.7305222 0.3942455 0.7706071 0.5722297 0.4774728 0.4827467 0.5304393
#> [31,] 0.8243869 0.6411725 0.9068725 0.8079242 0.8591547 0.7110088 0.6767730
#> [32,] 0.8269322 0.6286368 0.9104366 0.7955195 0.8268347 0.6997369 0.6824356
#> [33,] 0.8144045 0.6295397 0.8762760 0.7722860 0.7923069 0.6762234 0.6799983
#> [34,] 0.7650089 0.5460009 0.8448745 0.7176284 0.7319219 0.6415571 0.6467113
#> [35,] 0.7789569 0.5840524 0.8534509 0.7465641 0.6987643 0.6539010 0.6964318
#> [36,] 0.7729574 0.6386835 0.8487438 0.7298134 0.7719628 0.6664139 0.6419863
#> [37,] 0.7651225 0.6268309 0.8678487 0.7538796 0.7921902 0.6837304 0.6516047
#> [38,] 0.7959117 0.5354488 0.8914674 0.7529093 0.8029103 0.6529143 0.5833251
#> [39,] 0.7409542 0.4213132 0.7477081 0.6013373 0.5519989 0.5281832 0.5816178
#> [40,] 0.8118960 0.6140930 0.8942372 0.7853393 0.7680400 0.6879552 0.6998976
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5282071 0.8103250 0.8685127
#>  [2,] 0.4398597 0.6823326 0.8136275
#>  [3,] 0.5734181 0.8375300 0.8858189
#>  [4,] 0.5837901 0.8052727 0.8831037
#>  [5,] 0.3550633 0.5309138 0.7717191
#>  [6,] 0.4973049 0.7207070 0.8156181
#>  [7,] 0.5800646 0.8027522 0.8555561
#>  [8,] 0.4371918 0.6515852 0.8141305
#>  [9,] 0.5316895 0.7965163 0.8680069
#> [10,] 0.3087557 0.4787801 0.7316256
#> [11,] 0.6036676 0.8136213 0.8763491
#> [12,] 0.4780377 0.7046308 0.8164023
#> [13,] 0.4182485 0.7572733 0.8720144
#> [14,] 0.4686251 0.7276478 0.8392168
#> [15,] 0.5183567 0.7673688 0.8864153
#> [16,] 0.5509747 0.8373924 0.8855607
#> [17,] 0.6322048 0.8580529 0.9128766
#> [18,] 0.5025264 0.7915672 0.8478498
#> [19,] 0.5321544 0.7494877 0.8034037
#> [20,] 0.6104800 0.8530752 0.8762144
#> [21,] 0.3970545 0.6258346 0.8205680
#> [22,] 0.4635962 0.6191297 0.8278050
#> [23,] 0.5222497 0.7905142 0.8630276
#> [24,] 0.5728258 0.7854409 0.8416872
#> [25,] 0.4030574 0.6476475 0.8485184
#> [26,] 0.5766895 0.8403595 0.8787134
#> [27,] 0.3527561 0.5216406 0.7058004
#> [28,] 0.5766372 0.7722537 0.8112360
#> [29,] 0.3791098 0.5685137 0.6830474
#> [30,] 0.3435191 0.5452335 0.7918667
#> [31,] 0.5871088 0.8493384 0.8948428
#> [32,] 0.5738805 0.8381988 0.8975434
#> [33,] 0.5595569 0.8256009 0.8840700
#> [34,] 0.5141694 0.7171455 0.8269052
#> [35,] 0.5562069 0.7559910 0.8609891
#> [36,] 0.5787206 0.7930015 0.8417989
#> [37,] 0.5758834 0.7932174 0.8430698
#> [38,] 0.4821457 0.7651114 0.8585549
#> [39,] 0.3859936 0.6134902 0.7736368
#> [40,] 0.5912691 0.8174566 0.8891482
#> 
#> $post[[3]]
#>            [,1]      [,2]
#>  [1,] 0.7332959 0.2667041
#>  [2,] 0.7178734 0.2821266
#>  [3,] 0.7401576 0.2598424
#>  [4,] 0.7343507 0.2656493
#>  [5,] 0.6906925 0.3093075
#>  [6,] 0.7210074 0.2789926
#>  [7,] 0.7324148 0.2675852
#>  [8,] 0.7083829 0.2916171
#>  [9,] 0.7323307 0.2676693
#> [10,] 0.6815477 0.3184523
#> [11,] 0.7324370 0.2675630
#> [12,] 0.7194709 0.2805291
#> [13,] 0.7302560 0.2697440
#> [14,] 0.7212142 0.2787858
#> [15,] 0.7291585 0.2708415
#> [16,] 0.7414172 0.2585828
#> [17,] 0.7427474 0.2572526
#> [18,] 0.7336042 0.2663958
#> [19,] 0.7254602 0.2745398
#> [20,] 0.7422977 0.2577023
#> [21,] 0.7084072 0.2915928
#> [22,] 0.7026597 0.2973403
#> [23,] 0.7303386 0.2696614
#> [24,] 0.7277127 0.2722873
#> [25,] 0.7096549 0.2903451
#> [26,] 0.7388297 0.2611703
#> [27,] 0.6842730 0.3157270
#> [28,] 0.7281366 0.2718634
#> [29,] 0.6974509 0.3025491
#> [30,] 0.6931624 0.3068376
#> [31,] 0.7421565 0.2578435
#> [32,] 0.7402742 0.2597258
#> [33,] 0.7359560 0.2640440
#> [34,] 0.7237873 0.2762127
#> [35,] 0.7265208 0.2734792
#> [36,] 0.7287655 0.2712345
#> [37,] 0.7313823 0.2686177
#> [38,] 0.7311989 0.2688011
#> [39,] 0.7011385 0.2988615
#> [40,] 0.7353908 0.2646092
#> 
#> 
#> $pre
#> $pre[[1]]
#> NULL
#> 
#> $pre[[2]]
#>            [,1]        [,2]      [,3]      [,4]        [,5]          [,6]
#>  [1,] 1.2497549  0.43207664 1.8198733 1.2166231  1.33671485  0.6353538299
#>  [2,] 1.0374120  0.10352755 1.5623702 0.7882999  0.78577756  0.3777588743
#>  [3,] 1.4953176  0.54545386 2.1424391 1.3573479  1.72484985  0.8395744763
#>  [4,] 1.4435826  0.43676830 1.9686021 1.2408552  1.17199255  0.7690446136
#>  [5,] 0.8983208 -0.44780712 1.1848019 0.3578734 -0.29500910  0.0001939749
#>  [6,] 1.1997955 -0.06739094 1.7942819 0.8838057  1.08316807  0.4618294225
#>  [7,] 1.3388329  0.45727653 1.8665265 1.1382192  1.37349386  0.7176926433
#>  [8,] 1.0908866 -0.03948036 1.5332283 0.6655477  0.27068626  0.2844807896
#>  [9,] 1.3544054  0.48623389 1.8640455 1.1119452  1.22233599  0.6531794531
#> [10,] 0.7042842 -0.57888393 0.8952641 0.1834356 -0.23655924 -0.2672818097
#> [11,] 1.3200414  0.60886175 1.8470679 1.1813036  1.14660040  0.7418306703
#> [12,] 1.1454691 -0.15365450 1.8882086 0.9553807  0.99601892  0.4487030554
#> [13,] 1.4583459  0.07023039 1.9921765 1.0231344  1.19082678  0.4979718456
#> [14,] 1.1649112  0.25860819 1.5736130 0.8631959  0.79992258  0.4362863327
#> [15,] 1.3867061  0.22290423 2.2591653 1.2103935  0.76212589  0.6642582158
#> [16,] 1.5690617  0.45275580 2.3247551 1.3629350  1.84216725  0.8817603452
#> [17,] 1.6526627  0.59145596 2.3886976 1.5448572  1.49887430  0.9637204904
#> [18,] 1.3645204  0.23058742 2.0870870 1.1432993  1.61089286  0.6869433204
#> [19,] 1.1992265  0.13849424 1.7155497 0.9653098  1.28357889  0.6369495446
#> [20,] 1.5397270  0.50697578 2.2907259 1.4389516  1.97660103  1.0174632708
#> [21,] 1.1303418 -0.24602298 1.5985983 0.6066501  0.40681743  0.1467712458
#> [22,] 1.0746096 -0.01020029 1.3546909 0.5602498 -0.08390255  0.2129786893
#> [23,] 1.3783739  0.27549794 1.8573661 1.0654869  1.24778343  0.6129100107
#> [24,] 1.1646742  0.56365451 1.7569613 0.9882694  1.15172819  0.6660614602
#> [25,] 1.0723383 -0.06828139 1.5766334 0.7481656  0.13581303  0.2263182312
#> [26,] 1.4179222  0.60003632 2.0033463 1.2708525  1.60362374  0.8313661232
#> [27,] 0.5318603 -0.38911452 0.8534226 0.2022904  0.02224274 -0.2204458693
#> [28,] 1.1715506  0.22408880 1.7999740 1.0621980  1.52732067  0.6968484831
#> [29,] 0.6441698 -0.36433752 1.1690076 0.4543552  0.59816240  0.1150802739
#> [30,] 0.9972738 -0.42950109 1.2117425 0.2909541 -0.09016985 -0.0690406003
#> [31,] 1.5463567  0.58045661 2.2760321 1.4365781  1.80828701  0.9002886160
#> [32,] 1.5640395  0.52637329 2.3189779 1.3585225  1.56335805  0.8460454311
#> [33,] 1.4788873  0.53024281 1.9576276 1.2212646  1.33888710  0.7364695765
#> [34,] 1.1803398  0.18452537 1.6949533 0.9327277  1.00439553  0.5821286480
#> [35,] 1.2595983  0.33943123 1.7619275 1.0803705  0.84142065  0.6362312788
#> [36,] 1.2250861  0.56965472 1.7247821 0.9936761  1.21942782  0.6920101266
#> [37,] 1.1809720  0.51864494 1.8820698 1.1194119  1.33817827  0.7709688205
#> [38,] 1.3609354  0.14203341 2.1058185 1.1141894  1.40458383  0.6318740561
#> [39,] 1.0509338 -0.31738482 1.0864259 0.4110404  0.20875041  0.1128525012
#> [40,] 1.4623773  0.46454983 2.1347720 1.2970570  1.19727708  0.7905770927
#>               [,7]        [,8]        [,9]     [,10]
#>  [1,]  0.680381316  0.11294824  1.45212307 1.8878719
#>  [2,]  0.498166496 -0.24173156  0.76451227 1.4737553
#>  [3,]  0.721532886  0.29581069  1.63996383 2.0487267
#>  [4,]  0.908628442  0.33835169  1.41958078 2.0221555
#>  [5,]  0.257144067 -0.59685622  0.12381316 1.2180440
#>  [6,]  0.269963353 -0.01078051  0.94797112 1.4869369
#>  [7,]  0.659242499  0.32303871  1.40358517 1.7788602
#>  [8,]  0.322761352 -0.25256694  0.62601448 1.4770760
#>  [9,]  0.697226447  0.12692816  1.36466165 1.8834497
#> [10,] -0.030500221 -0.80594310 -0.08493079 1.0028859
#> [11,]  0.823375658  0.42077065  1.47371451 1.9583018
#> [12,]  0.146433250 -0.08790575  0.86944788 1.4921605
#> [13,]  0.424654614 -0.32996774  1.13778829 1.9188877
#> [14,]  0.516097657 -0.12566475  0.98272083 1.6524123
#> [15,]  0.726147506  0.07345981  1.19351314 2.0546372
#> [16,]  0.617764773  0.20460985  1.63895278 2.0461760
#> [17,]  1.022453997  0.54168717  1.79921118 2.3492750
#> [18,]  0.385117609  0.01010576  1.33439785 1.7178353
#> [19,]  0.424894352  0.12879521  1.09588204 1.4077051
#> [20,]  0.666409176  0.44933051  1.75892682 1.9570598
#> [21,]  0.270621957 -0.41775322  0.51438833 1.5202002
#> [22,]  0.577098988 -0.14587321  0.48585604 1.5701504
#> [23,]  0.505306902  0.08905771  1.32802787 1.8406668
#> [24,]  0.578526426  0.29338987  1.29765973 1.6708354
#> [25,]  0.504073934 -0.39274179  0.60871466 1.7230280
#> [26,]  0.793213367  0.30919811  1.66090510 1.9803027
#> [27,] -0.004538347 -0.60694646  0.08661648 0.8750738
#> [28,]  0.407173345  0.30898396  1.22108090 1.4580612
#> [29,] -0.104896917 -0.49332857  0.27578972 0.7678118
#> [30,]  0.121908129 -0.64765089  0.18142996 1.3362141
#> [31,]  0.738981182  0.35202592  1.72942175 2.1411922
#> [32,]  0.764987583  0.29770146  1.64488688 2.1702219
#> [33,]  0.753763819  0.23936380  1.55476535 2.0315493
#> [34,]  0.604614473  0.05669285  0.93034601 1.5638507
#> [35,]  0.830363728  0.22578182  1.13082438 1.8235297
#> [36,]  0.583995753  0.31752369  1.34311375 1.6716740
#> [37,]  0.626100402  0.30589697  1.34442906 1.6812485
#> [38,]  0.336438259 -0.07144741  1.18091021 1.8033392
#> [39,]  0.329417964 -0.46418462  0.46200701 1.2289618
#> [40,]  0.846810200  0.36921428  1.49920997 2.0820695
#> 
#> $pre[[3]]
#>            [,1]       [,2]
#>  [1,] 0.3855244 -0.6258853
#>  [2,] 0.3598677 -0.5740698
#>  [3,] 0.4012662 -0.6455217
#>  [4,] 0.3866535 -0.6301561
#>  [5,] 0.3092647 -0.4940941
#>  [6,] 0.3702846 -0.5791797
#>  [7,] 0.3841110 -0.6227978
#>  [8,] 0.3406970 -0.5468462
#>  [9,] 0.3840704 -0.6224097
#> [10,] 0.2929361 -0.4679573
#> [11,] 0.3789304 -0.6280918
#> [12,] 0.3686787 -0.5731599
#> [13,] 0.3900927 -0.6058291
#> [14,] 0.3630490 -0.5874433
#> [15,] 0.3801893 -0.6101680
#> [16,] 0.4078870 -0.6454605
#> [17,] 0.4032832 -0.6570146
#> [18,] 0.3952559 -0.6177304
#> [19,] 0.3767817 -0.5949281
#> [20,] 0.4075991 -0.6503462
#> [21,] 0.3455410 -0.5421197
#> [22,] 0.3234338 -0.5365615
#> [23,] 0.3833474 -0.6129936
#> [24,] 0.3721264 -0.6109223
#> [25,] 0.3412737 -0.5524347
#> [26,] 0.3966401 -0.6432546
#> [27,] 0.2949930 -0.4784862
#> [28,] 0.3801955 -0.6049932
#> [29,] 0.3286600 -0.5065287
#> [30,] 0.3153523 -0.4995934
#> [31,] 0.4052665 -0.6519408
#> [32,] 0.4018270 -0.6455671
#> [33,] 0.3907267 -0.6343277
#> [34,] 0.3705212 -0.5928053
#> [35,] 0.3701474 -0.6068942
#> [36,] 0.3745078 -0.6138606
#> [37,] 0.3815821 -0.6200649
#> [38,] 0.3908845 -0.6098291
#> [39,] 0.3292161 -0.5235090
#> [40,] 0.3890424 -0.6331058
#> 
#> 
#> $e
#>             [,1]       [,2]
#>  [1,]  0.2667041 -0.2667041
#>  [2,] -0.7178734  0.7178734
#>  [3,]  0.2598424 -0.2598424
#>  [4,] -0.7343507  0.7343507
#>  [5,] -0.6906925  0.6906925
#>  [6,]  0.2789926 -0.2789926
#>  [7,] -0.7324148  0.7324148
#>  [8,] -0.7083829  0.7083829
#>  [9,]  0.2676693 -0.2676693
#> [10,]  0.3184523 -0.3184523
#> [11,] -0.7324370  0.7324370
#> [12,] -0.7194709  0.7194709
#> [13,]  0.2697440 -0.2697440
#> [14,]  0.2787858 -0.2787858
#> [15,] -0.7291585  0.7291585
#> [16,]  0.2585828 -0.2585828
#> [17,] -0.7427474  0.7427474
#> [18,]  0.2663958 -0.2663958
#> [19,]  0.2745398 -0.2745398
#> [20,] -0.7422977  0.7422977
#> [21,]  0.2915928 -0.2915928
#> [22,] -0.7026597  0.7026597
#> [23,]  0.2696614 -0.2696614
#> [24,] -0.7277127  0.7277127
#> [25,]  0.2903451 -0.2903451
#> [26,]  0.2611703 -0.2611703
#> [27,]  0.3157270 -0.3157270
#> [28,] -0.7281366  0.7281366
#> [29,]  0.3025491 -0.3025491
#> [30,] -0.6931624  0.6931624
#> [31,]  0.2578435 -0.2578435
#> [32,]  0.2597258 -0.2597258
#> [33,]  0.2640440 -0.2640440
#> [34,] -0.7237873  0.7237873
#> [35,] -0.7265208  0.7265208
#> [36,] -0.7287655  0.7287655
#> [37,] -0.7313823  0.7313823
#> [38,]  0.2688011 -0.2688011
#> [39,] -0.7011385  0.7011385
#> [40,] -0.7353908  0.7353908
#> 
#> $L
#> [1] 0.6897950 0.7187174 0.7273743 0.7942508 0.8317088 0.8032211
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.4927536