Classification Imbalanced Random Forest Src Learner

Imbalanced Random forest for classification between two classes. Calls randomForestSRC::imbalanced.rfsrc() from from randomForestSRC.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.imbalanced_rfsrc")
lrn("classif.imbalanced_rfsrc")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, randomForestSRC

Parameters

Id	Type	Default	Levels	Range

ntree	integer	3000		\([1, \infty)\)

method	character	rfq	rfq, brf, standard	-

block.size	integer	10		\([1, \infty)\)

fast	logical	FALSE	TRUE, FALSE	-

ratio	numeric	-		\([0, 1]\)

mtry	integer	-		\([1, \infty)\)

mtry.ratio	numeric	-		\([0, 1]\)

nodesize	integer	15		\([1, \infty)\)

nodedepth	integer	-		\([1, \infty)\)

splitrule	character	gini	gini, auc, entropy	-

nsplit	integer	10		\([0, \infty)\)

importance	character	FALSE	FALSE, TRUE, none, permute, random, anti	-

bootstrap	character	by.root	by.root, by.node, none, by.user	-

samptype	character	swor	swor, swr	-

samp	untyped	-		-

membership	logical	FALSE	TRUE, FALSE	-

sampsize	untyped	-		-

sampsize.ratio	numeric	-		\([0, 1]\)

na.action	character	na.omit	na.omit, na.impute	-

nimpute	integer	1		\([1, \infty)\)

ntime	integer	-		\([1, \infty)\)

cause	integer	-		\([1, \infty)\)

proximity	character	FALSE	FALSE, TRUE, inbag, oob, all	-

distance	character	FALSE	FALSE, TRUE, inbag, oob, all	-

forest.wt	character	FALSE	FALSE, TRUE, inbag, oob, all	-

xvar.wt	untyped	-		-

split.wt	untyped	-		-

forest	logical	TRUE	TRUE, FALSE	-

var.used	character	FALSE	FALSE, all.trees, by.tree	-

split.depth	character	FALSE	FALSE, all.trees, by.tree	-

seed	integer	-		\((-\infty, -1]\)

do.trace	logical	FALSE	TRUE, FALSE	-

statistics	logical	FALSE	TRUE, FALSE	-

get.tree	untyped	-		-

outcome	character	train	train, test	-

ptn.count	integer	0		\([0, \infty)\)

cores	integer	1		\([1, \infty)\)

save.memory	logical	FALSE	TRUE, FALSE	-

perf.type	character	-	gmean, misclass, brier, none	-

Custom mlr3 parameters

mtry:

This hyperparameter can alternatively be set via the added hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

sampsize:

This hyperparameter can alternatively be set via the added hyperparameter sampsize.ratio as sampsize = max(ceiling(sampsize.ratio * n_obs), 1). Note that sampsize and sampsize.ratio are mutually exclusive.

cores: This value is set as the option rf.cores during training and is set to 1 by default.

References

O’Brien R, Ishwaran H (2019). “A random forests quantile classifier for class imbalanced data.” Pattern Recognition, 90, 232--249. doi:10.1016/j.patcog.2019.01.036 .

Chao C, Leo B (2004). “Using Random Forest to Learn Imbalanced Data.” University of California, Berkeley.

See also

Author

HarutyunyanLiana

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifImbalancedRandomForestSRC

Methods

Public methods

LearnerClassifImbalancedRandomForestSRC$new()

LearnerClassifImbalancedRandomForestSRC$importance()

LearnerClassifImbalancedRandomForestSRC$selected_features()

LearnerClassifImbalancedRandomForestSRC$oob_error()

LearnerClassifImbalancedRandomForestSRC$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifImbalancedRandomForestSRC$new()

Method importance()

The importance scores are extracted from the slot importance.

Usage

LearnerClassifImbalancedRandomForestSRC$importance()

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot var.used.

Usage

LearnerClassifImbalancedRandomForestSRC$selected_features()

Returns

character().

Method oob_error()

OOB error extracted from the model slot err.rate.

Usage

LearnerClassifImbalancedRandomForestSRC$oob_error()

Returns

numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifImbalancedRandomForestSRC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("classif.imbalanced_rfsrc")
print(learner)
#> <LearnerClassifImbalancedRandomForestSRC:classif.imbalanced_rfsrc>: Imbalanced Random Forest
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, randomForestSRC
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, oob_error, twoclass, weights

# available parameters:
learner$param_set$ids()
#>  [1] "ntree"          "method"         "block.size"     "fast"          
#>  [5] "ratio"          "mtry"           "mtry.ratio"     "nodesize"      
#>  [9] "nodedepth"      "splitrule"      "nsplit"         "importance"    
#> [13] "bootstrap"      "samptype"       "samp"           "membership"    
#> [17] "sampsize"       "sampsize.ratio" "na.action"      "nimpute"       
#> [21] "ntime"          "cause"          "proximity"      "distance"      
#> [25] "forest.wt"      "xvar.wt"        "split.wt"       "forest"        
#> [29] "var.used"       "split.depth"    "seed"           "do.trace"      
#> [33] "statistics"     "get.tree"       "outcome"        "ptn.count"     
#> [37] "cores"          "save.memory"    "perf.type"