Classification Light GBM Learner

Calls lightgbm::lgb.train from package lightgbm.

Details

For categorical features either pre-process data by encoding columns or specify the categorical columns with the categorical_feature parameter. For this learner please do not prefix the categorical feature with name:. Instead of providing the data that is used for early stopping explicitly, the parameter early_stopping_split determines the proportion of the training data that is used for early stopping.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.lightgbm")
lrn("classif.lightgbm")

Meta Information

Task type: “classif”

Predict Types: “prob”, “response”

Feature Types: “numeric”, “integer”

Required Packages: mlr3, mlr3extralearners, lightgbm

Parameters

Id	Type	Default	Levels	Range

nrounds	integer	5		\([1, \infty)\)

objective	character	binary	binary, multiclass, multiclassova, cross_entropy, cross_entropy_lambda, lambdarank, rank_xendcg	-

metric	character		, None, ndcg, map, auc, average_precision, binary_logloss, binary_error, auc_mu, multi_logloss, ...	-

custom_eval	untyped			-

verbose	integer	1		\((-\infty, \infty)\)

record	logical	TRUE	TRUE, FALSE	-

eval_freq	integer	1		\([1, \infty)\)

init_model	untyped			-

early_stopping_rounds	integer	-		\([1, \infty)\)

early_stopping_split	numeric	0		\([0, 1]\)

callbacks	untyped	-		-

reset_data	logical	FALSE	TRUE, FALSE	-

categorical_feature	untyped			-

boosting	character	gbdt	gbdt, rf, dart, goss	-

linear_tree	logical	FALSE	TRUE, FALSE	-

num_iterations	integer	100		\([0, \infty)\)

learning_rate	numeric	0.1		\([0, \infty)\)

num_leaves	integer	31		\([1, 131072]\)

tree_learner	character	serial	serial, feature, data, voting	-

num_threads	integer	0		\([0, \infty)\)

device_type	character	cpu	cpu, gpu	-

seed	integer	-		\((-\infty, \infty)\)

deterministic	logical	FALSE	TRUE, FALSE	-

force_col_wise	logical	FALSE	TRUE, FALSE	-

force_row_wise	logical	FALSE	TRUE, FALSE	-

histogram_pool_size	integer	-1		\((-\infty, \infty)\)

max_depth	integer	-1		\((-\infty, \infty)\)

min_data_in_leaf	integer	20		\([0, \infty)\)

min_sum_hessian_in_leaf	numeric	0.001		\([0, \infty)\)

bagging_fraction	numeric	1		\([0, 1]\)

pos_bagging_fraction	numeric	1		\([0, 1]\)

neg_bagging_fraction	numeric	1		\([0, 1]\)

bagging_freq	integer	0		\([0, \infty)\)

bagging_seed	integer	3		\((-\infty, \infty)\)

feature_fraction	numeric	1		\([0, 1]\)

feature_fraction_bynode	numeric	1		\([0, 1]\)

feature_fraction_seed	integer	2		\((-\infty, \infty)\)

extra_trees	logical	FALSE	TRUE, FALSE	-

extra_seed	integer	6		\((-\infty, \infty)\)

first_metric_only	logical	FALSE	TRUE, FALSE	-

max_delta_step	numeric	0		\((-\infty, \infty)\)

lambda_l1	numeric	0		\([0, \infty)\)

lambda_l2	numeric	0		\([0, \infty)\)

linear_lambda	numeric	0		\([0, \infty)\)

min_gain_to_split	numeric	0		\([0, \infty)\)

drop_rate	numeric	0.1		\([0, 1]\)

max_drop	integer	50		\((-\infty, \infty)\)

skip_drop	numeric	0.5		\([0, 1]\)

xgboost_dart_mode	logical	FALSE	TRUE, FALSE	-

uniform_drop	logical	FALSE	TRUE, FALSE	-

drop_seed	integer	4		\((-\infty, \infty)\)

top_rate	numeric	0.2		\([0, 1]\)

other_rate	numeric	0.1		\([0, 1]\)

min_data_per_group	integer	100		\([1, \infty)\)

max_cat_threshold	integer	32		\([1, \infty)\)

cat_l2	numeric	10		\([0, \infty)\)

cat_smooth	numeric	10		\([0, \infty)\)

max_cat_to_onehot	integer	4		\([1, \infty)\)

top_k	integer	20		\([1, \infty)\)

monotone_constraints	untyped			-

monotone_constraints_method	character	basic	basic, intermediate, advanced	-

monotone_penalty	numeric	0		\([0, \infty)\)

feature_contri	untyped			-

forcedsplits_filename	untyped			-

refit_decay_rate	numeric	0.9		\([0, 1]\)

cegb_tradeoff	numeric	1		\([0, \infty)\)

cegb_penalty_split	numeric	0		\([0, \infty)\)

cegb_penalty_feature_lazy	untyped	-		-

cegb_penalty_feature_coupled	untyped	-		-

path_smooth	numeric	0		\([0, \infty)\)

interaction_constraints	untyped	-		-

input_model	untyped			-

output_model	untyped	LightGBM_model.txt		-

saved_feature_importance_type	integer	0		\([0, 1]\)

snapshot_freq	integer	-1		\((-\infty, \infty)\)

max_bin	integer	255		\([2, \infty)\)

max_bin_by_feature	untyped			-

min_data_in_bin	integer	3		\([1, \infty)\)

bin_construct_sample_cnt	integer	200000		\([1, \infty)\)

data_random_seed	integer	1		\((-\infty, \infty)\)

is_enable_sparse	logical	TRUE	TRUE, FALSE	-

enable_bundle	logical	TRUE	TRUE, FALSE	-

use_missing	logical	TRUE	TRUE, FALSE	-

zero_as_missing	logical	FALSE	TRUE, FALSE	-

feature_pre_filter	logical	TRUE	TRUE, FALSE	-

pre_partition	logical	FALSE	TRUE, FALSE	-

two_round	logical	FALSE	TRUE, FALSE	-

header	logical	FALSE	TRUE, FALSE	-

group_column	untyped			-

forcedbins_filename	untyped			-

save_binary	logical	FALSE	TRUE, FALSE	-

objective_seed	integer	5		\((-\infty, \infty)\)

is_unbalance	logical	FALSE	TRUE, FALSE	-

scale_pos_weight	numeric	1		\([0, \infty)\)

sigmoid	numeric	1		\([0, \infty)\)

boost_from_average	logical	TRUE	TRUE, FALSE	-

lambdarank_truncation_level	integer	30		\([1, \infty)\)

lambdarank_norm	logical	TRUE	TRUE, FALSE	-

label_gain	untyped	-		-

metric_freq	integer	1		\([1, \infty)\)

is_provide_training_metric	logical	FALSE	TRUE, FALSE	-

eval_at	untyped	:, 1, 5		-

multi_error_top_k	integer	1		\([1, \infty)\)

auc_mu_weights	untyped			-

num_machines	integer	1		\([1, \infty)\)

local_listen_port	integer	12400		\([1, \infty)\)

time_out	integer	120		\([1, \infty)\)

machine_list_filename	untyped			-

machines	untyped			-

gpu_platform_id	integer	-1		\((-\infty, \infty)\)

gpu_device_id	integer	-1		\((-\infty, \infty)\)

gpu_use_dp	logical	FALSE	TRUE, FALSE	-

num_gpu	integer	1		\([1, \infty)\)

start_iteration	integer	0		\((-\infty, \infty)\)

num_iteration	integer	-1		\((-\infty, \infty)\)

pred_early_stop	logical	FALSE	TRUE, FALSE	-

pred_early_stop_freq	integer	10		\((-\infty, \infty)\)

pred_early_stop_margin	numeric	10		\((-\infty, \infty)\)

output_result	untyped	LightGBM_predict_result.txt		-

Custom mlr3 defaults

num_threads:

Actual default: 0L

Adjusted default: 1L

Reason for change: Prevents accidental conflicts with future.

verbose:

Actual default: 1L

Adjusted default: -1L

Reason for change: Prevents accidental conflicts with mlr messaging system.

References

Ke, Guolin, Meng, Qi, Finley, Thomas, Wang, Taifeng, Chen, Wei, Ma, Weidong, Ye, Qiwei, Liu, Tie-Yan (2017). “Lightgbm: A highly efficient gradient boosting decision tree.” Advances in neural information processing systems, 30.

See also

Author

kapsner

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLightGBM

Methods

Public methods

LearnerClassifLightGBM$new()

LearnerClassifLightGBM$importance()

LearnerClassifLightGBM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLightGBM$new()

Method importance()

The importance scores are extracted from lbg.importance.

Usage

LearnerClassifLightGBM$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLightGBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("lightgbm", quietly = TRUE)) {
  learner = mlr3::lrn("classif.lightgbm")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerClassifLightGBM:classif.lightgbm>
#> * Model: -
#> * Parameters: num_threads=1, verbose=-1
#> * Packages: mlr3, mlr3extralearners, lightgbm
#> * Predict Type: prob
#> * Feature types: numeric, integer
#> * Properties: importance, missings, multiclass, twoclass, weights
#>   [1] "nrounds"                       "objective"                    
#>   [3] "metric"                        "custom_eval"                  
#>   [5] "verbose"                       "record"                       
#>   [7] "eval_freq"                     "init_model"                   
#>   [9] "early_stopping_rounds"         "early_stopping_split"         
#>  [11] "callbacks"                     "reset_data"                   
#>  [13] "categorical_feature"           "boosting"                     
#>  [15] "linear_tree"                   "num_iterations"               
#>  [17] "learning_rate"                 "num_leaves"                   
#>  [19] "tree_learner"                  "num_threads"                  
#>  [21] "device_type"                   "seed"                         
#>  [23] "deterministic"                 "force_col_wise"               
#>  [25] "force_row_wise"                "histogram_pool_size"          
#>  [27] "max_depth"                     "min_data_in_leaf"             
#>  [29] "min_sum_hessian_in_leaf"       "bagging_fraction"             
#>  [31] "pos_bagging_fraction"          "neg_bagging_fraction"         
#>  [33] "bagging_freq"                  "bagging_seed"                 
#>  [35] "feature_fraction"              "feature_fraction_bynode"      
#>  [37] "feature_fraction_seed"         "extra_trees"                  
#>  [39] "extra_seed"                    "first_metric_only"            
#>  [41] "max_delta_step"                "lambda_l1"                    
#>  [43] "lambda_l2"                     "linear_lambda"                
#>  [45] "min_gain_to_split"             "drop_rate"                    
#>  [47] "max_drop"                      "skip_drop"                    
#>  [49] "xgboost_dart_mode"             "uniform_drop"                 
#>  [51] "drop_seed"                     "top_rate"                     
#>  [53] "other_rate"                    "min_data_per_group"           
#>  [55] "max_cat_threshold"             "cat_l2"                       
#>  [57] "cat_smooth"                    "max_cat_to_onehot"            
#>  [59] "top_k"                         "monotone_constraints"         
#>  [61] "monotone_constraints_method"   "monotone_penalty"             
#>  [63] "feature_contri"                "forcedsplits_filename"        
#>  [65] "refit_decay_rate"              "cegb_tradeoff"                
#>  [67] "cegb_penalty_split"            "cegb_penalty_feature_lazy"    
#>  [69] "cegb_penalty_feature_coupled"  "path_smooth"                  
#>  [71] "interaction_constraints"       "input_model"                  
#>  [73] "output_model"                  "saved_feature_importance_type"
#>  [75] "snapshot_freq"                 "max_bin"                      
#>  [77] "max_bin_by_feature"            "min_data_in_bin"              
#>  [79] "bin_construct_sample_cnt"      "data_random_seed"             
#>  [81] "is_enable_sparse"              "enable_bundle"                
#>  [83] "use_missing"                   "zero_as_missing"              
#>  [85] "feature_pre_filter"            "pre_partition"                
#>  [87] "two_round"                     "header"                       
#>  [89] "group_column"                  "forcedbins_filename"          
#>  [91] "save_binary"                   "objective_seed"               
#>  [93] "is_unbalance"                  "scale_pos_weight"             
#>  [95] "sigmoid"                       "boost_from_average"           
#>  [97] "lambdarank_truncation_level"   "lambdarank_norm"              
#>  [99] "label_gain"                    "metric_freq"                  
#> [101] "is_provide_training_metric"    "eval_at"                      
#> [103] "multi_error_top_k"             "auc_mu_weights"               
#> [105] "num_machines"                  "local_listen_port"            
#> [107] "time_out"                      "machine_list_filename"        
#> [109] "machines"                      "gpu_platform_id"              
#> [111] "gpu_device_id"                 "gpu_use_dp"                   
#> [113] "num_gpu"                       "start_iteration"              
#> [115] "num_iteration"                 "pred_early_stop"              
#> [117] "pred_early_stop_freq"          "pred_early_stop_margin"       
#> [119] "output_result"