Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLogistic$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

LearnerClassifLogistic$unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

LearnerClassifLogistic$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         944.7706
#> V10                        116.8527
#> V11                         52.8808
#> V12                        241.8027
#> V13                        -65.3714
#> V14                        -19.8762
#> V15                         55.0434
#> V16                        -53.8876
#> V17                          6.2962
#> V18                         87.5242
#> V19                        -74.6096
#> V2                           -86.75
#> V20                        139.2039
#> V21                       -143.1865
#> V22                         29.3181
#> V23                         40.1695
#> V24                         47.4801
#> V25                        -54.0631
#> V26                        -61.8719
#> V27                        121.2249
#> V28                        -23.7633
#> V29                        -31.8732
#> V3                        -630.6861
#> V30                        184.1728
#> V31                       -307.8117
#> V32                        132.4726
#> V33                         64.1525
#> V34                       -229.3842
#> V35                        198.1242
#> V36                         16.3081
#> V37                       -173.7561
#> V38                        -41.5291
#> V39                        200.4297
#> V4                          -10.769
#> V40                        -90.0648
#> V41                         26.6277
#> V42                        145.3307
#> V43                         -72.125
#> V44                          9.5267
#> V45                        -69.9986
#> V46                         -86.603
#> V47                        403.1714
#> V48                       -268.8371
#> V49                       1512.0782
#> V5                         -78.3119
#> V50                      -2680.8223
#> V51                        771.5054
#> V52                       -502.9294
#> V53                       2422.9866
#> V54                        514.0491
#> V55                      -1770.3886
#> V56                      -1109.4314
#> V57                      -2525.8799
#> V58                       2823.9499
#> V59                        1020.754
#> V6                         348.6383
#> V60                       -107.5349
#> V7                        -460.4083
#> V8                        -352.5133
#> V9                         238.4199
#> Intercept                 -126.7512
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         Infinity
#> V10           5.6035556436497455E50
#> V11            9.243306404799626E22
#> V12           1.031724335167036E105
#> V13                               0
#> V14                               0
#> V15            8.036034119689196E23
#> V16                               0
#> V17                        542.5146
#> V18           1.0263368866944014E38
#> V19                               0
#> V2                                0
#> V20            2.854347083568012E60
#> V21                               0
#> V22            5.403776946896946E12
#> V23          2.78866030524929824E17
#> V24            4.172110402387625E20
#> V25                               0
#> V26                               0
#> V27            4.439289423262467E52
#> V28                               0
#> V29                               0
#> V3                                0
#> V30            9.665413663168088E79
#> V31                               0
#> V32            3.404956866573087E57
#> V33            7.262637613879144E27
#> V34                               0
#> V35           1.1072517520605372E86
#> V36                   12091974.4569
#> V37                               0
#> V38                               0
#> V39           1.1104514557341803E87
#> V4                                0
#> V40                               0
#> V41            3.666531276224753E11
#> V42           1.3070825132452768E63
#> V43                               0
#> V44                      13720.9637
#> V45                               0
#> V46                               0
#> V47          1.2448621131059102E175
#> V48                               0
#> V49                        Infinity
#> V5                                0
#> V50                               0
#> V51                        Infinity
#> V52                               0
#> V53                        Infinity
#> V54          1.7728509906401746E223
#> V55                               0
#> V56                               0
#> V57                               0
#> V58                        Infinity
#> V59                        Infinity
#> V6           2.5803518591205713E151
#> V60                               0
#> V7                                0
#> V8                                0
#> V9            3.503104032837876E103
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768