Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         146.8761
#> V10                          3.1131
#> V11                        114.5976
#> V12                        144.0457
#> V13                       -241.4268
#> V14                         142.625
#> V15                        -27.0542
#> V16                        -46.8481
#> V17                       -101.2404
#> V18                        117.9931
#> V19                         25.3379
#> V2                        1036.9248
#> V20                         18.8622
#> V21                         -86.118
#> V22                        264.2417
#> V23                       -296.7285
#> V24                        302.8304
#> V25                       -119.5695
#> V26                          2.8047
#> V27                         82.4694
#> V28                        -17.3873
#> V29                         14.9561
#> V3                        -999.5516
#> V30                        197.8898
#> V31                       -413.3674
#> V32                        298.3693
#> V33                        -91.3815
#> V34                         54.6566
#> V35                         28.0674
#> V36                        -340.123
#> V37                         -0.9424
#> V38                        324.1338
#> V39                       -158.2345
#> V4                         611.5268
#> V40                        -95.1853
#> V41                        119.4291
#> V42                        -85.9244
#> V43                        286.1607
#> V44                       -110.5296
#> V45                         89.9266
#> V46                        -38.3218
#> V47                       -199.0708
#> V48                        364.3618
#> V49                        726.4038
#> V5                         -72.7283
#> V50                       -940.6763
#> V51                       2672.8804
#> V52                        2939.839
#> V53                       -1789.481
#> V54                      -1186.0751
#> V55                       1407.3754
#> V56                       -569.4184
#> V57                       -789.8871
#> V58                        -51.1539
#> V59                       5255.6728
#> V6                         462.8408
#> V60                      -3184.6705
#> V7                        -453.9937
#> V8                         -504.391
#> V9                         121.4681
#> Intercept                 -196.8731
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1             6.130229990092619E63
#> V10                         22.4897
#> V11            5.876040463345131E49
#> V12            3.616302458648863E62
#> V13                               0
#> V14            8.734423404678115E61
#> V15                               0
#> V16                               0
#> V17                               0
#> V18           1.7529481110807518E51
#> V19           1.0095263812155724E11
#> V2                         Infinity
#> V20                   155502765.215
#> V21                               0
#> V22           5.737207742085328E114
#> V23                               0
#> V24          3.2929075772226786E131
#> V25                               0
#> V26                         16.5226
#> V27            6.546225449927668E35
#> V28                               0
#> V29                    3128698.0021
#> V3                                0
#> V30            8.759210235252298E85
#> V31                               0
#> V32          3.8032939283723653E129
#> V33                               0
#> V34           5.4581384996478854E23
#> V35            1.547098259423163E12
#> V36                               0
#> V37                          0.3897
#> V38           5.882123153407903E140
#> V39                               0
#> V4           3.8256655541971465E265
#> V40                               0
#> V41            7.368992783757329E51
#> V42                               0
#> V43          1.8968032606110361E124
#> V44                               0
#> V45           1.1340039349923322E39
#> V46                               0
#> V47                               0
#> V48          1.7390784782821336E158
#> V49                        Infinity
#> V5                                0
#> V50                               0
#> V51                        Infinity
#> V52                        Infinity
#> V53                               0
#> V54                               0
#> V55                        Infinity
#> V56                               0
#> V57                               0
#> V58                               0
#> V59                        Infinity
#> V6           1.0213922693374544E201
#> V60                               0
#> V7                                0
#> V8                                0
#> V9              5.66160951615803E52
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623