Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                        -220.5699
#> V10                         -98.405
#> V11                        174.1086
#> V12                         57.9556
#> V13                          5.9164
#> V14                          37.749
#> V15                         66.3106
#> V16                         48.0353
#> V17                       -199.2469
#> V18                        182.3893
#> V19                       -149.2575
#> V2                         573.1269
#> V20                         13.8729
#> V21                         43.8341
#> V22                        -11.7207
#> V23                        -22.4622
#> V24                        161.1597
#> V25                         -2.6808
#> V26                       -187.3028
#> V27                        206.1108
#> V28                       -176.7071
#> V29                         14.2392
#> V3                       -1261.5623
#> V30                        286.2336
#> V31                       -369.7467
#> V32                        179.2802
#> V33                         12.4503
#> V34                         -99.529
#> V35                         147.679
#> V36                       -144.9743
#> V37                        -99.0002
#> V38                         19.3769
#> V39                        272.8107
#> V4                         846.6983
#> V40                       -415.9273
#> V41                        228.7283
#> V42                        -67.5064
#> V43                        -20.6803
#> V44                          90.132
#> V45                       -147.0627
#> V46                        229.9846
#> V47                        300.3031
#> V48                       -100.6428
#> V49                       1125.1683
#> V5                        -227.6326
#> V50                      -1782.8154
#> V51                       -783.0958
#> V52                       -201.0796
#> V53                       1011.6105
#> V54                         619.738
#> V55                      -3269.5288
#> V56                       1894.2569
#> V57                      -1087.8711
#> V58                       2168.9977
#> V59                       1333.1599
#> V6                         199.2775
#> V60                       -604.4881
#> V7                        -392.0907
#> V8                        -124.4495
#> V9                         109.0116
#> Intercept                  -62.2419
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                                0
#> V10                               0
#> V11            4.115144552298305E75
#> V12           1.4783693548978745E25
#> V13                        371.0862
#> V14           2.4783986761064576E16
#> V15           6.2850484921570715E28
#> V16            7.269196097009872E20
#> V17                               0
#> V18           1.6242505452679132E79
#> V19                               0
#> V2            8.051195862937065E248
#> V20                    1059094.8398
#> V21           1.0886898312192428E19
#> V22                               0
#> V23                               0
#> V24            9.789291235497486E69
#> V25                          0.0685
#> V26                               0
#> V27             3.25667870831144E89
#> V28                               0
#> V29                    1527557.7375
#> V3                                0
#> V30          2.0402416504895738E124
#> V31                               0
#> V32            7.251344566185654E77
#> V33                     255325.4587
#> V34                               0
#> V35           1.3683281482468844E64
#> V36                               0
#> V37                               0
#> V38                   260177503.735
#> V39           3.021091101706261E118
#> V4                         Infinity
#> V40                               0
#> V41           2.1649097379126116E99
#> V42                               0
#> V43                               0
#> V44            1.392562916856899E39
#> V45                               0
#> V46            7.603974939223404E99
#> V47          2.6301087090338985E130
#> V48                               0
#> V49                        Infinity
#> V5                                0
#> V50                               0
#> V51                               0
#> V52                               0
#> V53                        Infinity
#> V54           1.408554887003185E269
#> V55                               0
#> V56                        Infinity
#> V57                               0
#> V58                        Infinity
#> V59                        Infinity
#> V6            3.5086377007122343E86
#> V60                               0
#> V7                                0
#> V8                                0
#> V9             2.203675662323267E47
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478