Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         -91.9667
#> V10                        353.9545
#> V11                       -117.7987
#> V12                         66.0744
#> V13                        230.2188
#> V14                        -30.5925
#> V15                        -54.4667
#> V16                       -111.5797
#> V17                         73.5794
#> V18                        -96.3051
#> V19                        119.0778
#> V2                        1892.8668
#> V20                        -17.3599
#> V21                        -27.3434
#> V22                         21.6982
#> V23                         -20.901
#> V24                        247.8957
#> V25                       -184.5443
#> V26                        165.9631
#> V27                       -100.5533
#> V28                        -15.0318
#> V29                       -136.6135
#> V3                       -1506.4531
#> V30                        386.5111
#> V31                       -291.1971
#> V32                         86.8593
#> V33                         74.9947
#> V34                       -236.2921
#> V35                        250.9839
#> V36                       -169.4511
#> V37                        -67.6472
#> V38                        169.5209
#> V39                          7.8768
#> V4                          97.4383
#> V40                        -90.0944
#> V41                        -25.1781
#> V42                          2.5494
#> V43                        214.6195
#> V44                       -138.4474
#> V45                        -17.3575
#> V46                        -71.8934
#> V47                        253.2833
#> V48                        222.4868
#> V49                        467.5092
#> V5                         180.7119
#> V50                      -2577.8501
#> V51                       1065.0626
#> V52                       2708.6438
#> V53                       -430.2225
#> V54                       2380.9916
#> V55                      -4209.2143
#> V56                        515.6565
#> V57                       -381.7038
#> V58                      -1686.8891
#> V59                       2515.3006
#> V6                         140.8382
#> V60                      -3118.4222
#> V7                        -595.9995
#> V8                         -28.7165
#> V9                         -13.6216
#> Intercept                 -151.3435
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                                0
#> V10          5.2539145743450526E153
#> V11                               0
#> V12            4.963119328559216E28
#> V13            9.610402263013274E99
#> V14                               0
#> V15                               0
#> V16                               0
#> V17            9.018302766152298E31
#> V18                               0
#> V19             5.18603197214071E51
#> V2                         Infinity
#> V20                               0
#> V21                               0
#> V22                 2650991589.7285
#> V23                               0
#> V24          4.5680558917912485E107
#> V25                               0
#> V26           1.1935922408849023E72
#> V27                               0
#> V28                               0
#> V29                               0
#> V3                                0
#> V30           7.238340552100533E167
#> V31                               0
#> V32            5.278729712647778E37
#> V33            3.713539550531274E32
#> V34                               0
#> V35          1.0021655696625906E109
#> V36                               0
#> V37                               0
#> V38            4.187962739238488E73
#> V39                        2635.321
#> V4            2.0744540243950523E42
#> V40                               0
#> V41                               0
#> V42                         12.7993
#> V43            1.614641822724251E93
#> V44                               0
#> V45                               0
#> V46                               0
#> V47           9.989018150741464E109
#> V48           4.2148612393526423E96
#> V49          1.0881011809485737E203
#> V5             3.035196833485231E78
#> V50                               0
#> V51                        Infinity
#> V52                        Infinity
#> V53                               0
#> V54                        Infinity
#> V55                               0
#> V56           8.846350375602496E223
#> V57                               0
#> V58                               0
#> V59                        Infinity
#> V6             1.463070952659114E61
#> V60                               0
#> V7                                0
#> V8                                0
#> V9                                0
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551