Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.logistic")
print(learner)
#> <LearnerClassifLogistic:classif.logistic>: Multinomial Logistic Regression
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         145.8428
#> V10                         25.4976
#> V11                       -196.2562
#> V12                        540.5488
#> V13                       -326.0453
#> V14                        -46.8237
#> V15                        309.1823
#> V16                        -323.792
#> V17                       -145.4102
#> V18                        240.1251
#> V19                         39.5688
#> V2                          747.874
#> V20                         -58.241
#> V21                       -111.3434
#> V22                        191.4673
#> V23                        -53.8842
#> V24                        147.3517
#> V25                        -14.4795
#> V26                       -123.6405
#> V27                         47.9301
#> V28                          47.568
#> V29                        -101.995
#> V3                       -1145.1301
#> V30                        302.9991
#> V31                       -397.4992
#> V32                        189.9985
#> V33                          -7.453
#> V34                        -73.5937
#> V35                        156.3717
#> V36                       -222.5463
#> V37                        -36.0204
#> V38                          15.944
#> V39                         83.2407
#> V4                         260.9115
#> V40                       -123.0085
#> V41                         -9.6304
#> V42                        -53.6154
#> V43                         44.0189
#> V44                         -6.0955
#> V45                        165.1847
#> V46                         57.3292
#> V47                        226.5057
#> V48                        230.5914
#> V49                         55.9174
#> V5                         -66.5058
#> V50                      -2186.2568
#> V51                        618.0728
#> V52                       1679.1686
#> V53                       2285.8686
#> V54                        872.7989
#> V55                       -844.9864
#> V56                      -2606.6995
#> V57                        366.9541
#> V58                       -965.3459
#> V59                       6345.0659
#> V6                         119.4276
#> V60                      -1993.7984
#> V7                        -627.2478
#> V8                        -153.5846
#> V9                         575.0618
#> Intercept                 -109.6288
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1               2.1813199653395E63
#> V10           1.1842856649517871E11
#> V11                               0
#> V12           5.719268769028296E234
#> V13                               0
#> V14                               0
#> V15          1.8886509954150737E134
#> V16                               0
#> V17                               0
#> V18           1.927583122337909E104
#> V19          1.52942029382088576E17
#> V2                         Infinity
#> V20                               0
#> V21                               0
#> V22            1.422955351608164E83
#> V23                               0
#> V24            9.863094534691404E63
#> V25                               0
#> V26                               0
#> V27            6.542751808219242E20
#> V28            4.555480422831283E20
#> V29                               0
#> V3                                0
#> V30          3.8981267336410355E131
#> V31                               0
#> V32            3.275831002390406E82
#> V33                          0.0006
#> V34                               0
#> V35            8.153934231479687E67
#> V36                               0
#> V37                               0
#> V38                    8402501.1353
#> V39            1.415781616447403E36
#> V4           2.0531595249138745E113
#> V40                               0
#> V41                          0.0001
#> V42                               0
#> V43           1.3096547856220422E19
#> V44                          0.0023
#> V45            5.480453330551418E71
#> V46            7.902406178440853E24
#> V47           2.3452589747144427E98
#> V48          1.3949679530092579E100
#> V49           1.9258242873085164E24
#> V5                                0
#> V50                               0
#> V51          2.6644855240001504E268
#> V52                        Infinity
#> V53                        Infinity
#> V54                        Infinity
#> V55                               0
#> V56                               0
#> V57           2.323507387474939E159
#> V58                               0
#> V59                        Infinity
#> V6             7.358028350793456E51
#> V60                               0
#> V7                                0
#> V8                                0
#> V9           5.5740011615605185E249
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841