Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                        1255.3658
#> V10                       -161.2134
#> V11                        260.0046
#> V12                        112.4131
#> V13                        -169.981
#> V14                         95.7836
#> V15                        122.3326
#> V16                        -19.3278
#> V17                       -183.9286
#> V18                        118.6701
#> V19                         67.5586
#> V2                        -282.6594
#> V20                        -99.7119
#> V21                         45.1973
#> V22                         20.7348
#> V23                         10.6973
#> V24                        112.6914
#> V25                       -134.5369
#> V26                         -9.6249
#> V27                        159.2658
#> V28                         -86.706
#> V29                         99.0236
#> V3                        -307.5395
#> V30                         85.7979
#> V31                       -331.2359
#> V32                        127.3208
#> V33                         27.2656
#> V34                       -154.8877
#> V35                         221.048
#> V36                        -113.793
#> V37                       -172.9045
#> V38                         275.017
#> V39                         25.0807
#> V4                         484.9804
#> V40                       -248.8139
#> V41                         77.7601
#> V42                        119.6362
#> V43                        -45.9503
#> V44                        292.4335
#> V45                       -267.0853
#> V46                        118.9906
#> V47                         23.7295
#> V48                        534.7661
#> V49                        144.2061
#> V5                        -194.3069
#> V50                      -1045.6421
#> V51                       -657.1943
#> V52                        966.8813
#> V53                       2376.9435
#> V54                        880.1156
#> V55                       -550.3973
#> V56                      -2350.5456
#> V57                       -617.9492
#> V58                        275.9514
#> V59                        -34.4412
#> V6                         508.6216
#> V60                        -143.421
#> V7                        -531.4358
#> V8                        -141.5686
#> V9                         177.1359
#> Intercept                  -197.484
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         Infinity
#> V10                               0
#> V11           8.290243931474662E112
#> V12            6.612890130311914E48
#> V13                               0
#> V14            3.965351397924393E41
#> V15           1.3439350635493152E53
#> V16                               0
#> V17                               0
#> V18            3.449755043192974E51
#> V19            2.189447228366039E29
#> V2                                0
#> V20                               0
#> V21             4.25524776578712E19
#> V22                 1011636794.4914
#> V23                      44236.8493
#> V24            8.735215767919579E48
#> V25                               0
#> V26                          0.0001
#> V27           1.4732229216874218E69
#> V28                               0
#> V29           1.0125421094458176E43
#> V3                                0
#> V30           1.8261368609597612E37
#> V31                               0
#> V32           1.9711495034954518E55
#> V33            6.939243090063351E11
#> V34                               0
#> V35            9.998709484617757E95
#> V36                               0
#> V37                               0
#> V38          2.7438127173022105E119
#> V39            7.805827588055489E10
#> V4            4.210441177944341E210
#> V40                               0
#> V41             5.89891030077566E33
#> V42            9.064869583261694E51
#> V43                               0
#> V44           1.005223024509165E127
#> V45                               0
#> V46           4.7528973214133526E51
#> V47            2.021206126534309E10
#> V48          1.7617754480451156E232
#> V49            4.245455771579155E62
#> V5                                0
#> V50                               0
#> V51                               0
#> V52                        Infinity
#> V53                        Infinity
#> V54                        Infinity
#> V55                               0
#> V56                               0
#> V57                               0
#> V58           6.985356865140744E119
#> V59                               0
#> V6            7.790492047286908E220
#> V60                               0
#> V7                                0
#> V8                                0
#> V9             8.494221988205449E76
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3478261