Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         804.9712
#> V10                        122.9282
#> V11                        -304.443
#> V12                        308.6107
#> V13                        100.6925
#> V14                        -26.5745
#> V15                         88.5869
#> V16                       -218.6522
#> V17                         -6.3954
#> V18                         -1.0229
#> V19                         98.8922
#> V2                         745.0651
#> V20                        198.0647
#> V21                       -301.6899
#> V22                        142.7641
#> V23                         81.9825
#> V24                         23.4887
#> V25                         31.3335
#> V26                        100.5662
#> V27                       -161.2741
#> V28                         81.7668
#> V29                       -162.7283
#> V3                       -1107.2388
#> V30                        337.5779
#> V31                       -323.8675
#> V32                         99.9554
#> V33                         15.6665
#> V34                       -323.9332
#> V35                         373.267
#> V36                        -215.137
#> V37                       -102.8942
#> V38                         26.6282
#> V39                         76.0129
#> V4                         987.5628
#> V40                         -77.427
#> V41                        183.2578
#> V42                       -162.8871
#> V43                        249.2552
#> V44                         -236.35
#> V45                        382.4605
#> V46                       -111.7591
#> V47                       -117.7161
#> V48                        538.9202
#> V49                       -144.6354
#> V5                        -311.9327
#> V50                      -1717.8387
#> V51                       2105.7292
#> V52                      -2393.8865
#> V53                        601.1698
#> V54                        388.1567
#> V55                      -1915.2449
#> V56                      -1502.8106
#> V57                       1709.6418
#> V58                        495.7509
#> V59                        741.9887
#> V6                         290.2623
#> V60                      -2765.3048
#> V7                        -367.1207
#> V8                        -308.4889
#> V9                         371.6567
#> Intercept                 -137.0411
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         Infinity
#> V10           2.4381087080497117E53
#> V11                               0
#> V12          1.0664211375432507E134
#> V13            5.372536844069894E43
#> V14                               0
#> V15            2.970274256623979E38
#> V16                               0
#> V17                          0.0017
#> V18                          0.3596
#> V19             8.87815133540191E42
#> V2                         Infinity
#> V20           1.0433169779207201E86
#> V21                               0
#> V22            1.003809632201051E62
#> V23            4.022918707800718E35
#> V24           1.5886648961375341E10
#> V25           4.0547215827167734E13
#> V26            4.735217702369962E43
#> V27                               0
#> V28           3.2425676445281722E35
#> V29                               0
#> V3                                0
#> V30           4.057099909339025E146
#> V31                               0
#> V32           2.5708268555647827E43
#> V33                    6366330.8721
#> V34                               0
#> V35           1.281775569299449E162
#> V36                               0
#> V37                               0
#> V38           3.6685679019018414E11
#> V39           1.0279262076853611E33
#> V4                         Infinity
#> V40                               0
#> V41            3.871368614707185E79
#> V42                               0
#> V43          1.7789131093868261E108
#> V44                               0
#> V45          1.2603222170517269E166
#> V46                               0
#> V47                               0
#> V48             1.1221913555102E234
#> V49                               0
#> V5                                0
#> V50                               0
#> V51                        Infinity
#> V52                               0
#> V53          1.2154618770760626E261
#> V54           3.752517270380824E168
#> V55                               0
#> V56                               0
#> V57                        Infinity
#> V58          2.0038394357750082E215
#> V59                        Infinity
#> V6           1.1463362542131099E126
#> V60                               0
#> V7                                0
#> V8                                0
#> V9           2.5611999130406303E161
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478