Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         444.5711
#> V10                       -101.5724
#> V11                        324.0797
#> V12                       -136.7142
#> V13                         115.471
#> V14                        -38.7102
#> V15                         21.9877
#> V16                         -57.331
#> V17                          -57.63
#> V18                         29.2545
#> V19                         65.1738
#> V2                         -100.233
#> V20                         -1.5797
#> V21                         49.8745
#> V22                        -137.097
#> V23                         66.6342
#> V24                        133.2837
#> V25                       -127.2353
#> V26                         15.9488
#> V27                         11.8534
#> V28                          2.0394
#> V29                        -45.8113
#> V3                        -564.2488
#> V30                        118.7967
#> V31                        -135.451
#> V32                          10.219
#> V33                         47.2793
#> V34                       -174.4087
#> V35                        204.7834
#> V36                       -128.4895
#> V37                        -24.3164
#> V38                         -1.2488
#> V39                         78.7004
#> V4                         407.2704
#> V40                       -122.8325
#> V41                           8.651
#> V42                        -14.6286
#> V43                        165.3902
#> V44                         -35.496
#> V45                        -21.1935
#> V46                        193.0868
#> V47                         72.0485
#> V48                        140.4559
#> V49                         603.371
#> V5                        -151.7269
#> V50                      -2056.9117
#> V51                       -205.3531
#> V52                       -189.0087
#> V53                        378.4504
#> V54                        458.9096
#> V55                        173.8742
#> V56                       2676.2753
#> V57                      -1794.9521
#> V58                       1633.3942
#> V59                       -917.8598
#> V6                         162.3138
#> V60                      -2362.3707
#> V7                        -259.8174
#> V8                         -77.2819
#> V9                          60.5799
#> Intercept                  -33.3076
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1           1.1878803535832382E193
#> V10                               0
#> V11           5.572213718785666E140
#> V12                               0
#> V13           1.4074635834011373E50
#> V14                               0
#> V15                 3541017209.4142
#> V16                               0
#> V17                               0
#> V18               5.070837820902E12
#> V19            2.016695687884144E28
#> V2                                0
#> V20                           0.206
#> V21            4.572977939773256E21
#> V22                               0
#> V23            8.686853096796214E28
#> V24            7.662817429026425E57
#> V25                               0
#> V26                    8442680.5966
#> V27                     140558.4673
#> V28                          7.6863
#> V29                               0
#> V3                                0
#> V30            3.915305713906791E51
#> V31                               0
#> V32                      27418.5776
#> V33            3.413106653341974E20
#> V34                               0
#> V35            8.635755210066656E88
#> V36                               0
#> V37                               0
#> V38                          0.2868
#> V39           1.5105799891850846E34
#> V4            7.503867379298774E176
#> V40                               0
#> V41                       5715.6522
#> V42                               0
#> V43            6.730747983076536E71
#> V44                               0
#> V45                               0
#> V46            7.187033219702205E83
#> V47           1.9511266135006311E31
#> V48             9.98222035286104E60
#> V49          1.0982505053508154E262
#> V5                                0
#> V50                               0
#> V51                               0
#> V52                               0
#> V53          2.2851565322925806E164
#> V54          2.0040908007242567E199
#> V55           3.2555133404753036E75
#> V56                        Infinity
#> V57                               0
#> V58                        Infinity
#> V59                               0
#> V6            3.1045938325399306E70
#> V60                               0
#> V7                                0
#> V8                                0
#> V9             2.039501527096714E26
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3623188