Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         433.3992
#> V10                       -139.9089
#> V11                        332.7992
#> V12                         82.2038
#> V13                       -135.7674
#> V14                         81.8144
#> V15                         45.7345
#> V16                        -99.7768
#> V17                       -243.9519
#> V18                        193.1567
#> V19                        109.3688
#> V2                         797.1362
#> V20                         -2.7515
#> V21                       -220.9845
#> V22                        164.8448
#> V23                         -5.8414
#> V24                         26.1094
#> V25                         25.6929
#> V26                        -16.3122
#> V27                         70.8087
#> V28                        -98.3151
#> V29                          6.8765
#> V3                        -792.5417
#> V30                         42.7138
#> V31                       -159.3905
#> V32                         27.3502
#> V33                        139.2207
#> V34                       -153.6917
#> V35                          81.279
#> V36                       -109.8351
#> V37                        -68.6451
#> V38                        273.9501
#> V39                       -169.6971
#> V4                         123.7262
#> V40                       -118.4011
#> V41                        -44.9642
#> V42                          68.985
#> V43                         132.163
#> V44                          9.4448
#> V45                        104.6228
#> V46                       -130.9545
#> V47                        532.2492
#> V48                       -250.6534
#> V49                        205.5888
#> V5                        -117.6492
#> V50                      -1805.3641
#> V51                       1118.4707
#> V52                       4741.3204
#> V53                      -1471.6129
#> V54                        425.2101
#> V55                       1444.4769
#> V56                        2006.801
#> V57                        207.3714
#> V58                        185.4254
#> V59                       4052.7603
#> V6                         459.9809
#> V60                      -8591.4716
#> V7                        -600.2416
#> V8                          -97.961
#> V9                         221.5466
#> Intercept                  -95.4863
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1            1.670665330497792E188
#> V10                               0
#> V11          3.4109251893250054E144
#> V12            5.019399797010732E35
#> V13                               0
#> V14            3.400470325061198E35
#> V15            7.282036329349954E19
#> V16                               0
#> V17                               0
#> V18            7.706770785988834E83
#> V19           3.1495851030021758E47
#> V2                         Infinity
#> V20                          0.0638
#> V21                               0
#> V22           3.9009948024292105E71
#> V23                          0.0029
#> V24            2.183650319695776E11
#> V25           1.4397882528958218E11
#> V26                               0
#> V27            5.647299007571206E30
#> V28                               0
#> V29                        969.2674
#> V3                                0
#> V30            3.551134138132181E18
#> V31                               0
#> V32            7.551937668396162E11
#> V33           2.9024589875902724E60
#> V34                               0
#> V35           1.9907271765165307E35
#> V36                               0
#> V37                               0
#> V38           9.441402886006154E118
#> V39                               0
#> V4             5.414912957007521E53
#> V40                               0
#> V41                               0
#> V42             9.11632523072596E29
#> V43           2.4983207396627742E57
#> V44                      12641.6281
#> V45           2.7358266830683322E45
#> V46                               0
#> V47          1.4219150327050874E231
#> V48                               0
#> V49            1.932388112812327E89
#> V5                                0
#> V50                               0
#> V51                        Infinity
#> V52                        Infinity
#> V53                               0
#> V54          4.6387267756418544E184
#> V55                        Infinity
#> V56                        Infinity
#> V57           1.1488678975004875E90
#> V58           3.3822677564842656E80
#> V59                        Infinity
#> V6            5.849940037307823E199
#> V60                               0
#> V7                                0
#> V8                                0
#> V9            1.6461371148765217E96
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3768116