Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         216.7826
#> V10                          8.3307
#> V11                        -85.4271
#> V12                        309.1962
#> V13                        -33.6786
#> V14                       -180.7106
#> V15                        325.5837
#> V16                       -174.9055
#> V17                       -197.0616
#> V18                         11.9208
#> V19                        164.5334
#> V2                         325.6652
#> V20                        -23.8303
#> V21                         15.9018
#> V22                           0.521
#> V23                          44.311
#> V24                        159.7102
#> V25                       -173.4289
#> V26                        -66.2325
#> V27                        171.1507
#> V28                           -65.6
#> V29                        -60.1098
#> V3                        -779.4338
#> V30                        201.5339
#> V31                         -330.14
#> V32                        307.1769
#> V33                        -53.1923
#> V34                       -165.3927
#> V35                        205.3572
#> V36                       -121.8531
#> V37                        -37.3358
#> V38                         62.9168
#> V39                        -37.3673
#> V4                         586.7165
#> V40                        -25.7962
#> V41                        -37.0911
#> V42                         90.7988
#> V43                        -46.4079
#> V44                        -40.4842
#> V45                        -20.6439
#> V46                          6.5312
#> V47                        376.7857
#> V48                        216.4238
#> V49                        405.0626
#> V5                         -60.4069
#> V50                      -2060.2951
#> V51                        297.3201
#> V52                      -1881.5719
#> V53                       1392.0795
#> V54                        123.9882
#> V55                        -33.2457
#> V56                         -99.029
#> V57                       2549.3864
#> V58                        579.8369
#> V59                       2487.1728
#> V6                         158.1551
#> V60                      -2890.8678
#> V7                        -192.3269
#> V8                        -242.1481
#> V9                         223.0913
#> Intercept                 -122.8643
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1              1.40437855090638E94
#> V10                       4149.1923
#> V11                               0
#> V12          1.9150635602848115E134
#> V13                               0
#> V14                               0
#> V15          2.5073798086428373E141
#> V16                               0
#> V17                               0
#> V18                      150363.464
#> V19           2.8572942590513234E71
#> V2            2.720274918936121E141
#> V20                               0
#> V21                    8055108.7839
#> V22                          1.6837
#> V23           1.7538971091073108E19
#> V24           2.2974214864742434E69
#> V25                               0
#> V26                               0
#> V27           2.1368947009433613E74
#> V28                               0
#> V29                               0
#> V3                                0
#> V30           3.3502710560987494E87
#> V31                               0
#> V32          2.5424344211271105E133
#> V33                               0
#> V34                               0
#> V35           1.5328361798876757E89
#> V36                               0
#> V37                               0
#> V38           2.1105853262709829E27
#> V39                               0
#> V4            6.422961084569861E254
#> V40                               0
#> V41                               0
#> V42           2.7126865274569537E39
#> V43                               0
#> V44                               0
#> V45                               0
#> V46                        686.1887
#> V47           4.324444510307324E163
#> V48            9.809892865088934E93
#> V49           8.249900247596607E175
#> V5                                0
#> V50                               0
#> V51          1.3319345688190323E129
#> V52                               0
#> V53                        Infinity
#> V54            7.037199904501087E53
#> V55                               0
#> V56                               0
#> V57                        Infinity
#> V58           6.606505939297908E251
#> V59                        Infinity
#> V6              4.85148844020307E68
#> V60                               0
#> V7                                0
#> V8                                0
#> V9             7.714797083007573E96
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696