Classification Logistic Regression Learner

Multinomial Logistic Regression model with a ridge estimator. Calls RWeka::Logistic() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.logistic")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

C	logical	FALSE	TRUE, FALSE	-

R	numeric	-		\((-\infty, \infty)\)

M	integer	-1		\((-\infty, \infty)\)

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

le Cessie, S., van Houwelingen, J.C. (1992). “Ridge Estimators in Logistic Regression.” Applied Statistics, 41(1), 191-201.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogistic

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifLogistic$new()

LearnerClassifLogistic$marshal()

LearnerClassifLogistic$unmarshal()

LearnerClassifLogistic$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogistic$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifLogistic$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifLogistic$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogistic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.logistic")
print(learner)
#> 
#> ── <LearnerClassifLogistic> (classif.logistic): Multinomial Logistic Regression 
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Logistic Regression with ridge parameter of 1.0E-8
#> Coefficients...
#>                               Class
#> Variable                          M
#> ===================================
#> V1                         614.8246
#> V10                        -156.918
#> V11                        152.8143
#> V12                         260.227
#> V13                       -261.3948
#> V14                         26.3611
#> V15                        186.3102
#> V16                       -124.7529
#> V17                       -231.3309
#> V18                        232.4701
#> V19                         76.3157
#> V2                         143.6016
#> V20                       -123.7063
#> V21                         20.5492
#> V22                          56.893
#> V23                         11.6755
#> V24                         27.1471
#> V25                         54.5621
#> V26                       -193.4897
#> V27                        158.0793
#> V28                        -66.7434
#> V29                          0.9179
#> V3                        -892.1256
#> V30                        160.7743
#> V31                       -247.9963
#> V32                         55.2615
#> V33                        114.8655
#> V34                        -59.4915
#> V35                         79.9261
#> V36                       -152.2974
#> V37                         57.5201
#> V38                         -9.3248
#> V39                        107.5354
#> V4                         384.6749
#> V40                       -115.7764
#> V41                          27.881
#> V42                        -72.3829
#> V43                         27.0857
#> V44                        -56.7633
#> V45                          46.789
#> V46                        140.8208
#> V47                        211.0439
#> V48                         50.3469
#> V49                        690.8743
#> V5                        -177.4467
#> V50                      -2298.1264
#> V51                       -317.9496
#> V52                       1587.1315
#> V53                         1535.24
#> V54                        216.1058
#> V55                        536.2312
#> V56                       1554.2219
#> V57                        841.6783
#> V58                      -1352.3249
#> V59                       -342.7148
#> V6                         140.8152
#> V60                        -411.782
#> V7                         -79.5865
#> V8                         -50.9379
#> V9                         242.3745
#> Intercept                 -138.4985
#> 
#> 
#> Odds Ratios...
#>                               Class
#> Variable                          M
#> ===================================
#> V1           1.0350138480326537E267
#> V10                               0
#> V11           2.3248028220194374E66
#> V12          1.0355352133424436E113
#> V13                               0
#> V14            2.808609435886229E11
#> V15            8.193861453257388E80
#> V16                               0
#> V17                               0
#> V18           9.129775637550978E100
#> V19           1.3915386278062065E33
#> V2            2.3194270982985887E62
#> V20                               0
#> V21                  840240529.3199
#> V22            5.108634202416889E24
#> V23                     117655.2429
#> V24             6.16371458074715E11
#> V25           4.9660342447699707E23
#> V26                               0
#> V27            4.497659692301388E68
#> V28                               0
#> V29                          2.5041
#> V3                                0
#> V30            6.658871823497457E69
#> V31                               0
#> V32            9.994808557162691E23
#> V33            7.681977905327101E49
#> V34                               0
#> V35            5.145984971392191E34
#> V36                               0
#> V37            9.564272781951161E24
#> V38                          0.0001
#> V39            5.035294546889142E46
#> V4           1.1539931009281661E167
#> V40                               0
#> V41            1.283965766762339E12
#> V42                               0
#> V43             5.79669475015179E11
#> V44                               0
#> V45            2.090334413496178E20
#> V46            1.437784523571465E61
#> V47            4.520605307787441E91
#> V48            7.334859863474336E21
#> V49          1.1038058612146663E300
#> V5                                0
#> V50                               0
#> V51                               0
#> V52                        Infinity
#> V53                        Infinity
#> V54             7.13752649692516E93
#> V55           7.625164072501811E232
#> V56                        Infinity
#> V57                        Infinity
#> V58                               0
#> V59                               0
#> V6            1.4297348391513065E61
#> V60                               0
#> V7                                0
#> V8                                0
#> V9           1.8276882901033424E105
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406