Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1          V10        V11        V12        V13        V14
#>  [1,] 0.061026865  0.173964737 0.27325625 0.34784125 0.46320707 0.57181931
#>  [2,] 0.034696824  0.185297286 0.21366209 0.27239065 0.32148711 0.29036747
#>  [3,] 0.030930255  0.227465256 0.27379610 0.30538805 0.28779096 0.25507303
#>  [4,] 0.004792921  0.222282055 0.25004381 0.19643824 0.09069048 0.19167672
#>  [5,] 0.028553382  0.416645516 0.50584491 0.54963127 0.53442636 0.40345501
#>  [6,] 0.036229604  0.268914180 0.33581291 0.35022586 0.27763271 0.19043397
#>  [7,] 0.065154711  0.341286823 0.21303849 0.13861034 0.40240239 0.56625245
#>  [8,] 0.070700000  0.124700000 0.23400000 0.17640000 0.22840000 0.31150000
#>  [9,] 0.026800919  0.282175153 0.28584205 0.31206117 0.34417283 0.28721588
#> [10,] 0.026907909  0.606181231 0.66654924 0.53654792 0.46227852 0.30793643
#> [11,] 0.033219568  0.349595961 0.44497981 0.52128528 0.51373875 0.42116071
#> [12,] 0.025679447  0.198340978 0.21543353 0.23251127 0.28235107 0.37953701
#> [13,] 0.085198633  0.134620229 0.19818042 0.29726999 0.32255095 0.24005678
#> [14,] 0.023010500  0.137455757 0.09151615 0.02614041 0.14401007 0.28978791
#> [15,] 0.012981422  0.120587802 0.13765326 0.15750309 0.23515480 0.22853682
#> [16,] 0.020066943 -0.007026201 0.02570675 0.22468684 0.17289495 0.12704276
#> [17,] 0.011518851  0.062909475 0.06213850 0.06334202 0.07769559 0.09484420
#> [18,] 0.022373884  0.149868296 0.11264463 0.16478584 0.39524156 0.36451475
#> [19,] 0.019750666  0.081643865 0.06141696 0.06668623 0.10239715 0.01792829
#> [20,] 0.032366096  0.248228629 0.24703731 0.11585656 0.19662594 0.49208896
#> [21,] 0.023133447  0.333374996 0.28670473 0.16539046 0.26226884 0.49338315
#> [22,] 0.012390040  0.112146724 0.18126906 0.33245187 0.31974034 0.25050106
#> [23,] 0.025266766  0.064737282 0.02538943 0.04475747 0.08453066 0.08976677
#> [24,] 0.027099357  0.094496934 0.08562058 0.12128817 0.20822650 0.36898825
#> [25,] 0.017002098  0.122551483 0.14493463 0.13889039 0.19023112 0.21326295
#>               V15         V16        V17       V18       V19           V2
#>  [1,]  0.66396143 0.658677264 0.69661342 0.7948072 0.9057427  0.064282954
#>  [2,]  0.40394203 0.508735195 0.50497740 0.5361898 0.5738044  0.050985013
#>  [3,]  0.26596681 0.303297728 0.33442041 0.4472015 0.6115280  0.037620150
#>  [4,]  0.33498071 0.633920339 0.79839184 0.8179848 0.8055163  0.010952196
#>  [5,]  0.25614304 0.225516191 0.25385749 0.2751355 0.5131875  0.032114597
#>  [6,]  0.15962951 0.184045962 0.21599511 0.2160091 0.2499243  0.040111574
#>  [7,]  0.23881998 0.432156799 0.59444830 0.5774940 0.8173768  0.064569172
#>  [8,]  0.47250000 0.554300000 0.53860000 0.3746000 0.4583000  0.125200000
#>  [9,]  0.18244027 0.216864468 0.27876627 0.2526972 0.2568357  0.044473967
#> [10,]  0.20147640 0.245979700 0.26523384 0.3516991 0.5848761  0.039206487
#> [11,]  0.31263813 0.323683476 0.36344016 0.3456504 0.4537670  0.023163803
#> [12,]  0.46969503 0.534398348 0.61490220 0.6737927 0.7145112  0.015292329
#> [13,]  0.23696370 0.286604329 0.23344100 0.2889315 0.3951171  0.147558952
#> [14,]  0.57047755 0.820467632 0.85901470 0.8113210 0.8941866  0.016856553
#> [15,]  0.24717661 0.311383361 0.37490257 0.4042688 0.5643856  0.019940756
#> [16,]  0.08570508 0.129565540 0.18679251 0.1636443 0.1479477  0.029469207
#> [17,]  0.01303414 0.110328603 0.14881685 0.3564750 0.3415425  0.011943095
#> [18,]  0.16364835 0.100420878 0.02890613 0.3866652 0.7594115  0.036761156
#> [19,] -0.01159027 0.003602525 0.04742201 0.2092943 0.3616605 -0.002359654
#> [20,]  0.69208892 0.745020907 0.78337425 0.6240623 0.4215310  0.061893887
#> [21,]  0.61043725 0.740006946 0.78583315 0.7519975 0.6057605  0.039546945
#> [22,]  0.23012148 0.490544390 0.72665627 0.7240675 0.5373514  0.028923169
#> [23,]  0.13511211 0.067755518 0.10913701 0.2167282 0.3291456  0.061957470
#> [24,]  0.48964405 0.644218254 0.76132279 0.8170029 0.8165586  0.026886960
#> [25,]  0.23466625 0.242671938 0.21961802 0.2342546 0.2376755  0.020460207
#>              V20       V21        V22       V23       V24        V25       V26
#>  [1,] 0.87572321 0.8482830 0.85623363 0.8820380 0.8215266 0.61507481 0.5012357
#>  [2,] 0.67544274 0.7327790 0.83123226 0.9658175 1.0059267 0.98096764 1.0181096
#>  [3,] 0.70428577 0.8059694 0.82211538 0.8305759 0.9128853 0.93043465 1.0049434
#>  [4,] 0.69475652 0.6948853 0.80315872 0.8164202 0.4253882 0.10843435 0.2622365
#>  [5,] 0.71399641 0.8295520 0.86642641 0.7378030 0.7393769 0.72604382 0.7761570
#>  [6,] 0.26767964 0.3992366 0.53826217 0.6438581 0.6914591 0.69540361 0.7616965
#>  [7,] 0.88284911 0.5264151 0.63850095 0.8381033 0.6714584 0.77667576 0.8746006
#>  [8,] 0.59610000 0.7464000 0.76440000 0.5711000 0.6257000 0.66950000 0.7131000
#>  [9,] 0.30757427 0.2814670 0.39714398 0.5578124 0.5842978 0.62706088 0.6892996
#> [10,] 0.73381444 0.8596515 0.82986567 0.7715442 0.8051145 0.74516832 0.7883109
#> [11,] 0.61903149 0.7185892 0.85823791 0.8826888 0.8929379 0.94687169 0.9769185
#> [12,] 0.75302975 0.8301727 0.89664052 0.9300271 0.9520179 0.91875396 0.8765725
#> [13,] 0.22101696 0.2528955 0.56159585 0.8612070 0.9086586 0.74534882 0.5746646
#> [14,] 0.89794077 0.9858882 0.98382476 0.9760692 0.9818457 0.91811303 0.9624232
#> [15,] 0.62605049 0.5855839 0.48800978 0.4922950 0.6020371 0.72637227 0.7792310
#> [16,] 0.06891821 0.1257309 0.06717279 0.2051182 0.2085824 0.11479386 0.2693634
#> [17,] 0.26873820 0.1630694 0.16813758 0.2001921 0.2445286 0.08214014 0.1995654
#> [18,] 1.01359860 0.9877870 0.88387427 0.6620153 0.4165799 0.32389302 0.7441177
#> [19,] 0.41965971 0.4304296 0.51702473 0.6630234 0.7759941 0.88649620 0.8798646
#> [20,] 0.30235787 0.5157575 0.75140984 0.8521218 0.9898624 0.86298844 0.6802138
#> [21,] 0.49231358 0.5341756 0.66105680 0.7593731 0.8308738 0.75630311 0.7072401
#> [22,] 0.38009735 0.4107916 0.46153317 0.5803189 0.7363283 0.85305946 0.8905353
#> [23,] 0.30162413 0.1750221 0.21900860 0.3490468 0.4358584 0.47624699 0.2594329
#> [24,] 0.84136873 0.9832110 0.94738186 0.8791241 0.8776083 0.89007531 0.9198263
#> [25,] 0.33700560 0.4231383 0.50141397 0.6113612 0.7160008 0.74516771 0.7754550
#>              V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.49105110 0.4632492 0.3753388 0.072576483 0.2914148 0.2365545 0.2545948
#>  [2,] 0.97465202 0.9725690 0.8165686 0.011384032 0.6007458 0.5336779 0.4753466
#>  [3,] 0.98879052 0.9547585 0.8645797 0.034357884 0.6795949 0.4891086 0.2977093
#>  [4,] 0.58391452 0.7432606 0.8306851 0.001567731 0.7434664 0.4054596 0.4265332
#>  [5,] 0.91124811 0.9626164 0.8527651 0.045831579 0.5894049 0.4680375 0.4785545
#>  [6,] 0.89857610 0.9641794 0.9121587 0.034539828 0.7852279 0.6762257 0.6118629
#>  [7,] 0.82811303 0.6879368 0.3374742 0.073375403 0.6400292 0.7138969 0.8203966
#>  [8,] 0.75670000 0.8077000 0.8477000 0.144700000 0.9289000 0.9513000 0.7995000
#>  [9,] 0.74605421 0.8812742 0.8719756 0.042856967 0.8698523 0.9151127 0.8993869
#> [10,] 0.83102568 0.8800879 0.7239675 0.052280254 0.4548120 0.3397093 0.2970802
#> [11,] 0.96308408 0.7410110 0.4566029 0.020995647 0.3650045 0.3483746 0.4315235
#> [12,] 0.85650244 0.7967514 0.6860514 0.034686084 0.5580855 0.4088693 0.3073961
#> [13,] 0.47586740 0.5485026 0.7148632 0.195478513 0.5027152 0.4812670 0.3835669
#> [14,] 0.95971086 0.7557032 0.4430875 0.043210282 0.3683046 0.4590930 0.4152205
#> [15,] 0.81210546 0.7699969 0.5781124 0.028143837 0.4222619 0.3680810 0.3018231
#> [16,] 0.52283066 0.6727699 0.6808068 0.034521184 0.6433633 0.7772085 0.6544474
#> [17,] 0.43814709 0.7028598 0.8639208 0.013734230 0.7754845 0.5635602 0.2165624
#> [18,] 0.97912778 0.7416587 0.4881629 0.049092666 0.1572574 0.3099679 0.4181383
#> [19,] 0.76401237 0.8151200 0.9708597 0.008684750 0.9631215 0.9199942 0.8950164
#> [20,] 0.69075349 0.5848731 0.4666341 0.098359826 0.5180472 0.4342212 0.1571791
#> [21,] 0.70534556 0.5341167 0.3594553 0.056340568 0.3737856 0.4537161 0.4079596
#> [22,] 0.91453003 0.8990372 0.9204565 0.027401414 0.8685391 0.7647957 0.6912902
#> [23,] 0.08390328 0.2872024 0.5654992 0.036759838 0.8430689 0.8342983 0.7436084
#> [24,] 0.79606807 0.5587187 0.3690837 0.032037951 0.4056286 0.4332780 0.2860590
#> [25,] 0.74678998 0.7696337 0.7688286 0.027955725 0.6655488 0.6270779 0.5645254
#>             V33        V34        V35        V36        V37        V38
#>  [1,] 0.2425726 0.28717951 0.36896915 0.24274196 0.28456156 0.34619476
#>  [2,] 0.3317017 0.18110253 0.14570179 0.19843905 0.32732310 0.36637852
#>  [3,] 0.2180599 0.19962435 0.18904346 0.09210737 0.08306986 0.13092253
#>  [4,] 0.3966270 0.38494051 0.26580308 0.16679176 0.29384157 0.51039818
#>  [5,] 0.4406409 0.30540603 0.10540532 0.18827417 0.21320762 0.29302064
#>  [6,] 0.4876696 0.30665271 0.15175618 0.08131791 0.17737205 0.13660449
#>  [7,] 0.6672680 0.52736684 0.29810337 0.07266028 0.09732630 0.39285769
#>  [8,] 0.4362000 0.40480000 0.49520000 0.17120000 0.36520000 0.37630000
#>  [9,] 0.8046040 0.67710218 0.59576834 0.48981386 0.39048942 0.36690419
#> [10,] 0.3446360 0.41438313 0.30060969 0.37107007 0.27788367 0.35205931
#> [11,] 0.4033338 0.30324767 0.08254939 0.27742449 0.39663666 0.32872173
#> [12,] 0.2121579 0.11636767 0.13396705 0.16109779 0.22017310 0.22565486
#> [13,] 0.3474619 0.18085041 0.07786640 0.42934211 0.51278440 0.26678056
#> [14,] 0.3537264 0.31306503 0.22789761 0.08808494 0.26556988 0.19762288
#> [15,] 0.2734597 0.27621373 0.28489488 0.29821912 0.32210808 0.32675446
#> [16,] 0.6687039 0.83648858 0.92875535 0.96008095 0.78805835 0.66715066
#> [17,] 0.4804281 0.47324422 0.20630518 0.30686457 0.46233253 0.55035314
#> [18,] 0.3209805 0.32739686 0.34122022 0.22072418 0.25107478 0.27517672
#> [19,] 1.0005407 1.00579889 0.80864467 0.50949003 0.21975561 0.15034042
#> [20,] 0.1795222 0.08032619 0.41469811 0.70830316 0.72588194 0.35398552
#> [21,] 0.3374246 0.32975508 0.55170734 0.64978640 0.59804267 0.34556163
#> [22,] 0.7391426 0.85700956 0.97319746 0.99335134 0.93472184 0.81039571
#> [23,] 0.8174033 0.95671508 1.00646635 0.89794367 0.54062509 0.27556149
#> [24,] 0.2548111 0.25961293 0.16338077 0.06867884 0.12377284 0.09745946
#> [25,] 0.5319629 0.57432144 0.59236828 0.54307368 0.38527903 0.25397600
#>              V39            V4       V40        V41        V42        V43
#>  [1,] 0.35553690  0.0773744701 0.2924028 0.26546607 0.32282855 0.28340205
#>  [2,] 0.31196330  0.0261642508 0.3065385 0.30475882 0.24496348 0.21074546
#>  [3,] 0.12206420  0.0387314249 0.1469349 0.08580585 0.11784684 0.21493738
#>  [4,] 0.46036244  0.0071114318 0.3344671 0.13647261 0.23839229 0.46851454
#>  [5,] 0.55023803  0.0272837719 0.5146389 0.18013223 0.30899630 0.36804232
#>  [6,] 0.12337392  0.0571566471 0.2176933 0.33657426 0.26407182 0.24298500
#>  [7,] 0.60142854  0.1128346160 0.3530557 0.39310981 0.48017347 0.45886196
#>  [8,] 0.28410000  0.1644000000 0.0427000 0.53310000 0.69520000 0.42880000
#>  [9,] 0.33250794  0.0534208129 0.4211050 0.55418154 0.52110164 0.40475181
#> [10,] 0.53742481  0.0704039716 0.4073013 0.16764067 0.33059700 0.40297229
#> [11,] 0.32634651  0.0185274816 0.2934508 0.11629332 0.16713711 0.35631718
#> [12,] 0.20237567  0.0476958826 0.2342699 0.27127154 0.24541567 0.20410898
#> [13,] 0.24017402  0.2071412663 0.3030842 0.29962363 0.40987937 0.28609825
#> [14,] 0.09249993  0.0391678495 0.1163692 0.07303012 0.03646691 0.08838646
#> [15,] 0.34021231  0.0268610555 0.2599680 0.18166092 0.17144842 0.13514456
#> [16,] 0.61817012  0.0219290153 0.6373589 0.59271312 0.52365129 0.39202555
#> [17,] 0.61226905  0.0274883959 0.4785123 0.23981050 0.19718822 0.13585765
#> [18,] 0.16667470  0.0459161041 0.2822108 0.25064657 0.17269317 0.21830167
#> [19,] 0.10415434 -0.0008421084 0.2877159 0.36187074 0.40417720 0.31907599
#> [20,] 0.08350040  0.0821725813 0.3074916 0.42603476 0.39363394 0.25144662
#> [21,] 0.33228279  0.0587235627 0.4508599 0.39649479 0.34721172 0.24330045
#> [22,] 0.63919650  0.0334618330 0.4597604 0.29147665 0.18954709 0.15022894
#> [23,] 0.20326132  0.0164293486 0.2207681 0.33419076 0.39581864 0.15937308
#> [24,] 0.07706499  0.0419682753 0.1658494 0.11903207 0.07723912 0.05076519
#> [25,] 0.22364119  0.0319234829 0.2895324 0.24798883 0.23361902 0.24154052
#>              V44          V45          V46        V47        V48        V49
#>  [1,] 0.19486064  0.136287463  0.097858599 0.04475731 0.02280711 0.03850294
#>  [2,] 0.20676719  0.152230628  0.130572752 0.09769267 0.05962625 0.03652091
#>  [3,] 0.20993607  0.210551276  0.186547721 0.13729150 0.07668056 0.05239845
#>  [4,] 0.35680873  0.085057343  0.173582713 0.15125863 0.07915629 0.05831085
#>  [5,] 0.45221405  0.476733513  0.201269905 0.08931752 0.13482956 0.08286690
#>  [6,] 0.15318886  0.147591978  0.143181671 0.15243947 0.15389126 0.08958540
#>  [7,] 0.41450340  0.540943207  0.562144608 0.31828602 0.24757859 0.12959693
#>  [8,] 0.30630000  0.583500000  0.569200000 0.26300000 0.11960000 0.09830000
#>  [9,] 0.41276120  0.507672660  0.431177551 0.27075318 0.17824312 0.09629668
#> [10,] 0.39255432  0.288784987  0.088976008 0.11494980 0.16246098 0.09340319
#> [11,] 0.44075775  0.415453873  0.262230397 0.13175254 0.05623055 0.02448450
#> [12,] 0.17559154  0.123633625  0.099895847 0.12251776 0.08249116 0.05181512
#> [13,] 0.10448151  0.152855820  0.215728259 0.29665980 0.26446130 0.13058248
#> [14,] 0.14116401  0.130112093  0.043403258 0.05171513 0.04584222 0.02091738
#> [15,] 0.10575984  0.082194349  0.073547064 0.04620838 0.03531833 0.01640076
#> [16,] 0.22790451  0.112638098  0.056734749 0.08399705 0.08910418 0.05229772
#> [17,] 0.05031847 -0.004628592 -0.007198245 0.06460449 0.06394527 0.01563853
#> [18,] 0.11070349  0.092021265  0.117268763 0.14065883 0.11668630 0.07708122
#> [19,] 0.26919221  0.247656478  0.181555567 0.08434780 0.02672487 0.02398615
#> [20,] 0.19915623  0.230066188  0.198902377 0.19298131 0.17492685 0.12548809
#> [21,] 0.19897077  0.167410689  0.138648684 0.11678065 0.09749377 0.05984747
#> [22,] 0.10142522  0.171565966  0.191471563 0.11079403 0.05826283 0.01652292
#> [23,] 0.22060690  0.141106996  0.130127280 0.16007741 0.10955855 0.05422923
#> [24,] 0.11748633  0.108266121  0.073485180 0.11526762 0.08127278 0.03503612
#> [25,] 0.19839396  0.097951752  0.068556587 0.05190562 0.04803139 0.01365597
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.10260881 0.017159216 0.018928253 0.019772634 0.014070504 0.025822066
#>  [2,] 0.04254526 0.005121467 0.016301700 0.003955532 0.011613878 0.011710156
#>  [3,] 0.09986046 0.016556755 0.022315586 0.013178602 0.008870235 0.003234014
#>  [4,] 0.02581566 0.016896577 0.009741686 0.019004747 0.010307111 0.003748673
#>  [5,] 0.08567704 0.010544942 0.027230200 0.019700927 0.005428776 0.006833757
#>  [6,] 0.05575973 0.017684761 0.017089546 0.013137428 0.007239271 0.011704937
#>  [7,] 0.24708799 0.082155010 0.024256947 0.020931455 0.035924389 0.023773107
#>  [8,] 0.16930000 0.037400000 0.029100000 0.015600000 0.019700000 0.013500000
#>  [9,] 0.07363369 0.039223306 0.022834285 0.024644772 0.014797619 0.010128610
#> [10,] 0.10348568 0.027186851 0.027705668 0.015904338 0.017037405 0.016103521
#> [11,] 0.04339085 0.010496311 0.007849616 0.005114345 0.009617247 0.009116353
#> [12,] 0.09304063 0.026214451 0.017341598 0.018506320 0.017334749 0.011279009
#> [13,] 0.19766287 0.020725815 0.031838454 0.042695382 0.022584239 0.026172828
#> [14,] 0.06930608 0.007360391 0.016293338 0.009965460 0.013417162 0.007687473
#> [15,] 0.03727798 0.014253045 0.013375684 0.014415748 0.011528885 0.012741039
#> [16,] 0.04869736 0.015107022 0.008469647 0.006950859 0.018028066 0.011972203
#> [17,] 0.04474161 0.006778641 0.004174133 0.019512132 0.026977522 0.015558068
#> [18,] 0.06427055 0.043793791 0.005592045 0.015017191 0.013160186 0.014292200
#> [19,] 0.00507954 0.009480919 0.006003266 0.010714184 0.007051674 0.006732097
#> [20,] 0.04369993 0.032499824 0.016506821 0.010690083 0.005598842 0.012421502
#> [21,] 0.07375727 0.018746990 0.016579555 0.013949278 0.006411672 0.012220607
#> [22,] 0.02740244 0.024657270 0.024700641 0.017933038 0.007910090 0.007940321
#> [23,] 0.04874832 0.011539872 0.016727039 0.003198802 0.005138819 0.008290588
#> [24,] 0.07014813 0.022567264 0.007789726 0.013388432 0.011027766 0.003496309
#> [25,] 0.06825375 0.010520978 0.010468747 0.006040456 0.008020231 0.007441980
#>               V55          V56           V57         V58         V59
#>  [1,] 0.016475111 0.0107337265  0.0080421829 0.019073993 0.014192680
#>  [2,] 0.011646087 0.0098058417  0.0054747870 0.004221817 0.009098991
#>  [3,] 0.005890921 0.0040602268  0.0045147445 0.006386737 0.004173373
#>  [4,] 0.005192983 0.0075851582 -0.0006515714 0.005133419 0.002746689
#>  [5,] 0.009544391 0.0034417525  0.0035921021 0.005227451 0.005736649
#>  [6,] 0.006353901 0.0101551182  0.0032965394 0.008728655 0.010760255
#>  [7,] 0.044489267 0.0392096604  0.0353079610 0.043790400 0.024204264
#>  [8,] 0.012700000 0.0138000000  0.0133000000 0.013100000 0.015400000
#>  [9,] 0.006952684 0.0055027550  0.0081627070 0.006080513 0.003199044
#> [10,] 0.009401025 0.0062006405  0.0048928024 0.006863272 0.007690941
#> [11,] 0.005668187 0.0075749350  0.0045156967 0.004313937 0.008067721
#> [12,] 0.004064153 0.0095254629  0.0109070653 0.007454822 0.008044917
#> [13,] 0.014910733 0.0130233716  0.0162240499 0.018007202 0.020097147
#> [14,] 0.002401014 0.0074874729  0.0045961622 0.003273135 0.004489500
#> [15,] 0.013083678 0.0079723053  0.0086274912 0.009641116 0.003962181
#> [16,] 0.007358248 0.0079453964  0.0087338468 0.009508456 0.013099688
#> [17,] 0.019257711 0.0052989856  0.0039900970 0.004355983 0.009690661
#> [18,] 0.005592200 0.0045675204  0.0066062088 0.009313853 0.005849205
#> [19,] 0.002685621 0.0001758625  0.0005707217 0.002220386 0.005820132
#> [20,] 0.006356757 0.0054884083  0.0115887149 0.007167835 0.007515274
#> [21,] 0.009804583 0.0041389368  0.0086858288 0.006520805 0.009448149
#> [22,] 0.009415555 0.0075741264  0.0080829177 0.010030155 0.006585841
#> [23,] 0.007385852 0.0073352822  0.0100999710 0.016089418 0.009627068
#> [24,] 0.005207163 0.0131267941  0.0105893898 0.001704436 0.005862501
#> [25,] 0.005554963 0.0082827101  0.0098698554 0.005259753 0.008212059
#>                V6          V60          V7           V8          V9
#>  [1,] 0.145892295 0.0089000478 0.178714696  0.130941398  0.15238799
#>  [2,] 0.074681235 0.0038887963 0.091599909  0.118805160  0.14117220
#>  [3,] 0.118979990 0.0047587410 0.116189942  0.151155773  0.18855885
#>  [4,] 0.019072369 0.0026429272 0.008464169 -0.008054319  0.11656308
#>  [5,] 0.155402954 0.0049691188 0.111820787  0.179068726  0.27388851
#>  [6,] 0.103675250 0.0074857008 0.101303785  0.112178630  0.19051743
#>  [7,] 0.125468873 0.0097660108 0.179142010  0.099378115  0.24586858
#>  [8,] 0.084400000 0.0218000000 0.071500000  0.094700000  0.15830000
#>  [9,] 0.132914918 0.0019020508 0.125245816  0.142740932  0.28367705
#> [10,] 0.162901519 0.0058591273 0.203336992  0.266267763  0.39719045
#> [11,] 0.096990376 0.0049498652 0.111549939  0.125908308  0.29663650
#> [12,] 0.122159332 0.0070745130 0.145763876  0.163814822  0.18842341
#> [13,] 0.256648519 0.0093171172 0.240906803  0.136754235  0.07701289
#> [14,] 0.093119913 0.0031212165 0.132566836  0.098096090  0.13678928
#> [15,] 0.055706968 0.0045998971 0.076204470  0.056708874  0.09131301
#> [16,] 0.080912076 0.0063824946 0.047083968  0.020884256 -0.02691517
#> [17,] 0.053822877 0.0093688967 0.016097622  0.044396760  0.09908015
#> [18,] 0.060332945 0.0018444315 0.074538068  0.005311525  0.06438716
#> [19,] 0.001815657 0.0043381457 0.043639397  0.001739525  0.01834376
#> [20,] 0.068344451 0.0040259690 0.159474119  0.219758993  0.23187066
#> [21,] 0.139136922 0.0071492855 0.110591862  0.170380343  0.24774566
#> [22,] 0.101567664 0.0104060404 0.143327458  0.083354587  0.06964272
#> [23,] 0.079477493 0.0007470332 0.092598338  0.020459200  0.05661053
#> [24,] 0.077938097 0.0047671463 0.113857630  0.122015553  0.12222776
#> [25,] 0.121509031 0.0064764351 0.115079050  0.119354232  0.12394311
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913