Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12       V13        V14
#>  [1,] 0.03553255 0.32826854 0.29004082 0.32159909 0.3764120 0.37275493
#>  [2,] 0.10210000 0.13690000 0.25090000 0.26310000 0.2796000 0.29770000
#>  [3,] 0.04418389 0.34038149 0.38893715 0.52278178 0.6075486 0.76557217
#>  [4,] 0.02956165 0.31276094 0.43402454 0.48660008 0.4554657 0.33203452
#>  [5,] 0.01720306 0.31036825 0.33441274 0.29485200 0.3053711 0.27092047
#>  [6,] 0.03287525 0.13756846 0.11200543 0.15119394 0.2771814 0.31004741
#>  [7,] 0.02844832 0.26096900 0.33666591 0.35228286 0.2693400 0.14553947
#>  [8,] 0.02252778 0.22469131 0.25276030 0.29548062 0.3231491 0.34980253
#>  [9,] 0.03944929 0.01125607 0.11623556 0.21501919 0.2351000 0.28591535
#> [10,] 0.03491508 0.23699215 0.34924930 0.46881794 0.5025431 0.48264614
#> [11,] 0.05543491 0.26697448 0.31801922 0.32214116 0.4975704 0.74615478
#> [12,] 0.01141443 0.35854671 0.39036619 0.34282875 0.2335224 0.14802215
#> [13,] 0.02136287 0.11471972 0.09095151 0.05211626 0.1619596 0.30157677
#> [14,] 0.01086588 0.09876743 0.09631282 0.04902088 0.1493281 0.33889009
#> [15,] 0.06073536 0.13174458 0.23959846 0.38636896 0.3439811 0.15982273
#> [16,] 0.01787339 0.10971845 0.09937924 0.12910663 0.2480498 0.27482386
#> [17,] 0.02297517 0.05017126 0.01653216 0.20182761 0.1645954 0.08709542
#> [18,] 0.02530000 0.21970000 0.26530000 0.32230000 0.5582000 0.69160000
#> [19,] 0.01895555 0.12754950 0.15286368 0.15377966 0.1786123 0.18710088
#> [20,] 0.02593658 0.17362568 0.21852738 0.09788140 0.1815151 0.38669669
#> [21,] 0.01941971 0.15319471 0.17594965 0.18437477 0.2192182 0.19768418
#> [22,] 0.03759604 0.23744024 0.24344298 0.11834788 0.1878753 0.48677269
#> [23,] 0.02620000 0.61940000 0.63330000 0.70600000 0.5544000 0.53200000
#> [24,] 0.02930000 0.08600000 0.04140000 0.04720000 0.0835000 0.09380000
#> [25,] 0.02278201 0.37678626 0.27681800 0.08792748 0.1939288 0.49422579
#>             V15       V16       V17       V18        V19          V2        V20
#>  [1,] 0.3154586 0.3633281 0.3429718 0.2466256 0.42327783 0.046309087 0.52474925
#>  [2,] 0.3823000 0.3129000 0.3956000 0.2093000 0.32180000 0.083000000 0.33450000
#>  [3,] 0.8255578 0.8427469 0.9806523 0.9011290 0.92093634 0.036067362 0.83133043
#>  [4,] 0.2395198 0.2997414 0.3412387 0.3390190 0.45683938 0.043879985 0.62641778
#>  [5,] 0.1798445 0.1392701 0.2656106 0.4304978 0.63990219 0.045456438 0.78180752
#>  [6,] 0.3195511 0.3356177 0.3528311 0.2124421 0.20881084 0.061113039 0.28965442
#>  [7,] 0.1609527 0.2160505 0.2605756 0.2484883 0.27800450 0.045865730 0.28624012
#>  [8,] 0.3774508 0.4344996 0.4962374 0.5805091 0.68359237 0.029142207 0.74702057
#>  [9,] 0.4018794 0.5265927 0.6872185 0.6963711 0.83727887 0.053885463 0.94343092
#> [10,] 0.5597718 0.7124434 0.8463840 0.9339744 0.97787844 0.066275896 0.94139085
#> [11,] 0.8632524 0.8443829 0.7409491 0.8330348 0.93941898 0.068252120 0.91252767
#> [12,] 0.2003462 0.3283420 0.3554830 0.4239522 0.52955446 0.027431630 0.60383998
#> [13,] 0.5446077 0.7613567 0.8059759 0.7834119 0.85726268 0.017413090 0.88027682
#> [14,] 0.4791673 0.6569062 0.7766449 0.7839583 0.81432562 0.009402367 0.81092091
#> [15,] 0.2524215 0.3707822 0.5521156 0.6442806 0.58241678 0.054385870 0.34205773
#> [16,] 0.2400971 0.2700756 0.2580228 0.3639060 0.64294987 0.034828411 0.80900094
#> [17,] 0.1240845 0.1889656 0.1848114 0.1157422 0.09696957 0.033577031 0.08677188
#> [18,] 0.7943000 0.7152000 0.3512000 0.2008000 0.26760000 0.080800000 0.42990000
#> [19,] 0.2008776 0.2467920 0.2867608 0.3099543 0.32849369 0.020594172 0.36953373
#> [20,] 0.4974864 0.6980649 0.8810720 0.9206103 0.76386406 0.025128021 0.48293305
#> [21,] 0.2591114 0.3446115 0.3956439 0.3800722 0.31844408 0.020801083 0.35219859
#> [22,] 0.6969786 0.7449025 0.7789766 0.6145563 0.41480969 0.061704919 0.29876615
#> [23,] 0.6479000 0.6931000 0.6759000 0.7551000 0.89290000 0.058200000 0.86190000
#> [24,] 0.1466000 0.0809000 0.1179000 0.2179000 0.33260000 0.064400000 0.32580000
#> [25,] 0.6306747 0.7151833 0.7583206 0.6718061 0.44038405 0.045676070 0.35194017
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.5366490 0.47687610 0.5264463 0.5723062 0.62135446 0.7201450 0.7567659
#>  [2,] 0.3184000 0.28870000 0.3610000 0.2566000 0.41060000 0.4591000 0.4722000
#>  [3,] 0.8239343 0.83501306 0.6089681 0.2465381 0.18012139 0.1976241 0.3349672
#>  [4,] 0.7581837 0.86153559 0.8453897 0.8275660 0.85106971 0.8742572 0.9183662
#>  [5,] 0.7138600 0.43368385 0.2327519 0.2705996 0.47286779 0.8024020 0.9726395
#>  [6,] 0.3118647 0.45832095 0.5442784 0.5123983 0.50766262 0.5153108 0.5608231
#>  [7,] 0.4242530 0.55412387 0.6462791 0.7240397 0.74928213 0.8284955 0.9351780
#>  [8,] 0.8210614 0.86578262 0.8838596 0.9301184 0.91869081 0.9224730 0.9158063
#>  [9,] 0.9179042 0.92936970 0.8696569 0.8190499 0.67932924 0.5516748 0.5167080
#> [10,] 0.8270892 0.64667690 0.5421869 0.4736149 0.30535292 0.1199588 0.0811846
#> [11,] 0.9049634 0.92915970 1.0097125 0.9265198 0.70016795 0.6665154 0.5529316
#> [12,] 0.6985891 0.85313113 0.9209360 0.9603983 0.91053234 0.7805055 0.5445854
#> [13,] 0.9819986 0.97674578 0.9583883 0.9668126 0.91028027 0.9371810 0.9109043
#> [14,] 0.8240783 0.84794233 0.8877038 0.9000248 0.91469008 0.9849213 1.0130774
#> [15,] 0.2348776 0.20677576 0.2195538 0.2092379 0.29509012 0.2927001 0.2797964
#> [16,] 0.7525472 0.64348333 0.4699265 0.4433447 0.59924609 0.7576060 0.8663873
#> [17,] 0.1878840 0.07301445 0.1811882 0.2097369 0.07785791 0.2081270 0.5454425
#> [18,] 0.5280000 0.34890000 0.1430000 0.5453000 0.63380000 0.7712000 0.6838000
#> [19,] 0.3920351 0.46147545 0.5749236 0.6824983 0.70563151 0.7112079 0.7160204
#> [20,] 0.5297290 0.69750816 0.7824007 0.8285118 0.72484966 0.6416133 0.5673455
#> [21,] 0.4256747 0.46065168 0.5808122 0.7072754 0.78551980 0.7812764 0.7473620
#> [22,] 0.5265166 0.75945793 0.8555621 0.9837395 0.87257332 0.6721667 0.6756603
#> [23,] 0.7974000 0.67370000 0.4293000 0.3648000 0.53310000 0.2413000 0.5070000
#> [24,] 0.2111000 0.23020000 0.3361000 0.4259000 0.46090000 0.2606000 0.0874000
#> [25,] 0.4884273 0.69690437 0.8249217 0.8881461 0.76780548 0.7178747 0.7439599
#>              V28       V29         V3       V30          V31        V32
#>  [1,] 0.80549023 0.7392782 0.05343219 0.8024622  0.807408412 0.70760499
#>  [2,] 0.72780000 0.7591000 0.05770000 0.6579000  0.751400000 0.66660000
#>  [3,] 0.30402285 0.1126958 0.01251422 0.1090065 -0.002220685 0.07805013
#>  [4,] 0.85070765 0.6576871 0.04184429 0.4783033  0.364795192 0.43132259
#>  [5,] 0.90920454 0.8369193 0.05116640 0.7839375  0.565063280 0.50810079
#>  [6,] 0.74661022 0.7940396 0.09757092 0.7681756  0.809242575 0.81155543
#>  [7,] 0.93216300 0.7893116 0.05182382 0.7203646  0.658361701 0.57987108
#>  [8,] 0.86797534 0.7678377 0.03655775 0.5984381  0.415509708 0.30473000
#>  [9,] 0.40066042 0.3838821 0.08864374 0.3938247  0.227462474 0.21179586
#> [10,] 0.05679197 0.1031907 0.07507088 0.2127844  0.265926015 0.32377687
#> [11,] 0.47517339 0.5008438 0.06029648 0.3017579  0.267359583 0.15803536
#> [12,] 0.34918041 0.1568742 0.02536416 0.2797465  0.379551737 0.34750005
#> [13,] 0.70445794 0.4336455 0.03438480 0.3895897  0.448551694 0.38524424
#> [14,] 0.91615964 0.6700111 0.01142608 0.4603423  0.289265409 0.29675732
#> [15,] 0.45049135 0.7276692 0.06779482 0.7433229  0.821744836 0.79249188
#> [16,] 0.73914414 0.5344926 0.03339168 0.4269826  0.366542514 0.29748779
#> [17,] 0.77210848 0.7853084 0.01246855 0.7114627  0.794698061 0.63044648
#> [18,] 0.80150000 0.8073000 0.05070000 0.8310000  0.779200000 0.50490000
#> [19,] 0.81218128 0.8788579 0.02510095 0.8053493  0.699374416 0.62721130
#> [20,] 0.42388021 0.3673109 0.06183825 0.4404162  0.254197768 0.18312571
#> [21,] 0.76359370 0.6471496 0.02587675 0.4600308  0.367283171 0.31100037
#> [22,] 0.59195242 0.4686674 0.09426817 0.5200846  0.433039558 0.16331795
#> [23,] 0.85330000 0.6036000 0.10990000 0.8514000  0.851200000 0.50450000
#> [24,] 0.28620000 0.5606000 0.03900000 0.8344000  0.809600000 0.72500000
#> [25,] 0.60146736 0.3919125 0.08262914 0.4402498  0.534968002 0.36269339
#>              V33        V34       V35         V36       V37       V38
#>  [1,] 0.48373592 0.34669383 0.1968661  0.14933300 0.1937063 0.2984026
#>  [2,] 0.49030000 0.59620000 0.6552000  0.40140000 0.1188000 0.3245000
#>  [3,] 0.23396832 0.09653152 0.2210316 -0.02734607 0.1838930 0.2742929
#>  [4,] 0.40628627 0.35562907 0.2493013  0.37696839 0.3919281 0.3952711
#>  [5,] 0.68782399 0.64273817 0.5231478  0.37839289 0.3977697 0.4104960
#>  [6,] 0.76850560 0.55385564 0.4968529  0.38180992 0.1989509 0.3091549
#>  [7,] 0.44101172 0.35517750 0.2296824  0.07174587 0.1914318 0.2018835
#>  [8,] 0.23656442 0.16142812 0.1542940  0.12143655 0.1451577 0.1920051
#>  [9,] 0.28652088 0.25252286 0.3122543  0.33076463 0.3833428 0.4308824
#> [10,] 0.32907015 0.44827480 0.4483344  0.30986334 0.3419174 0.4006218
#> [11,] 0.01935976 0.10260607 0.2267635  0.23101519 0.2105788 0.1556891
#> [12,] 0.46696666 0.59756425 0.5183583  0.31729994 0.1843956 0.1645337
#> [13,] 0.34493619 0.30086098 0.1999457  0.08061700 0.2201900 0.1718093
#> [14,] 0.16974183 0.21213410 0.2616899  0.23323445 0.2202785 0.1633629
#> [15,] 0.65120069 0.64340220 0.6969144  0.78361234 0.8663968 0.9650200
#> [16,] 0.32393407 0.29022616 0.2791779  0.31869684 0.4264688 0.4362283
#> [17,] 0.65226364 0.87370417 0.9927614  0.94225242 0.7245197 0.5818035
#> [18,] 0.14130000 0.27670000 0.5084000  0.47870000 0.1356000 0.2299000
#> [19,] 0.66742538 0.71131713 0.6635130  0.57444732 0.4596706 0.3439269
#> [20,] 0.17867349 0.10263927 0.3711199  0.53698281 0.4309331 0.2245869
#> [21,] 0.22704784 0.21359880 0.2758024  0.29479920 0.2265127 0.1647635
#> [22,] 0.18834090 0.08834822 0.4212916  0.71541476 0.7381743 0.3738628
#> [23,] 0.18620000 0.27090000 0.4232000  0.30430000 0.6116000 0.6756000
#> [24,] 0.80480000 0.94350000 1.0000000  0.89600000 0.5516000 0.3037000
#> [25,] 0.22389183 0.19069474 0.5099383  0.64888014 0.6047619 0.3290778
#>              V39         V4         V40        V41        V42        V43
#>  [1,] 0.41656211 0.04002950 0.422459701 0.54708986 0.56406896 0.55325488
#>  [2,] 0.31070000 0.06270000 0.135400000 0.51090000 0.79880000 0.75170000
#>  [3,] 0.20168926 0.12628486 0.070274239 0.05886117 0.32077625 0.26875857
#>  [4,] 0.50895944 0.03635949 0.441386051 0.26469225 0.26953425 0.38204044
#>  [5,] 0.35425760 0.07363167 0.288654870 0.38458794 0.38839265 0.34024537
#>  [6,] 0.26306278 0.09879579 0.213229994 0.56566310 0.60015116 0.38934941
#>  [7,] 0.15197412 0.06051172 0.271764412 0.35242509 0.23444838 0.25120487
#>  [8,] 0.18865802 0.04546299 0.195627421 0.19701139 0.20336052 0.18024320
#>  [9,] 0.36469169 0.10231845 0.361524183 0.32253326 0.37623638 0.35856283
#> [10,] 0.24804676 0.07693128 0.268632737 0.40407255 0.30098269 0.15695736
#> [11,] 0.31529914 0.05154778 0.287216063 0.17181771 0.21267824 0.35259920
#> [12,] 0.23504729 0.02811144 0.331715972 0.26222254 0.26584009 0.23373678
#> [13,] 0.10697869 0.03918214 0.132779900 0.08942452 0.04107222 0.07749612
#> [14,] 0.03486925 0.03692562 0.005600634 0.04099452 0.02053813 0.05783180
#> [15,] 0.91560572 0.02452776 0.791490199 0.74309146 0.65947343 0.54579429
#> [16,] 0.44234373 0.03671756 0.329558135 0.21161292 0.18062919 0.10977928
#> [17,] 0.59499983 0.01147650 0.667003342 0.55581942 0.50869666 0.36237860
#> [18,] 0.27890000 0.02440000 0.383300000 0.29330000 0.11550000 0.17050000
#> [19,] 0.26410746 0.03477250 0.262237519 0.22842376 0.22983828 0.22401826
#> [20,] 0.11662818 0.06182988 0.309361294 0.38930337 0.29040896 0.16878073
#> [21,] 0.17155167 0.03307920 0.156562283 0.19551956 0.17259161 0.15477343
#> [22,] 0.07863549 0.07865023 0.294183807 0.42237140 0.39509504 0.25726348
#> [23,] 0.53750000 0.10830000 0.471900000 0.46470000 0.25870000 0.21290000
#> [24,] 0.23380000 0.01730000 0.238200000 0.33180000 0.38210000 0.15750000
#> [25,] 0.31995125 0.06789972 0.474153338 0.40705761 0.31622231 0.23892911
#>               V44        V45          V46        V47        V48        V49
#>  [1,]  0.48563776 0.61589914  0.544802770 0.37116627 0.26153029 0.15237843
#>  [2,]  0.55080000 0.58580000  0.729200000 0.55220000 0.33390000 0.16080000
#>  [3,] -0.02148185 0.22778428  0.175367422 0.03771964 0.06020158 0.05545138
#>  [4,]  0.39485213 0.42418588  0.263483566 0.13070198 0.09551734 0.06702474
#>  [5,]  0.20347097 0.10974343  0.084389413 0.07502743 0.06202717 0.02831056
#>  [6,]  0.48944707 0.58578744  0.521684198 0.33701436 0.14725125 0.07381573
#>  [7,]  0.20930588 0.17608505  0.129703013 0.13169452 0.14197155 0.09247348
#>  [8,]  0.18604484 0.17758971  0.148664075 0.11908433 0.07940872 0.04751163
#>  [9,]  0.22577669 0.13796004  0.104011409 0.09237939 0.06374258 0.03434104
#> [10,]  0.10176979 0.05227058  0.054352769 0.07143056 0.04674659 0.03687891
#> [11,]  0.24381646 0.10228618  0.116663017 0.10164282 0.08620778 0.07335807
#> [12,]  0.16177681 0.16686495  0.139391143 0.14951073 0.13414105 0.07514651
#> [13,]  0.13228993 0.11696038  0.042545412 0.05462732 0.04615758 0.02432572
#> [14,]  0.03759777 0.02987080 -0.005015443 0.02935175 0.02509859 0.01767537
#> [15,]  0.42057547 0.22470250  0.212572675 0.18957730 0.13556373 0.07603710
#> [16,]  0.07776542 0.09338430  0.089321750 0.07227687 0.04541486 0.02524168
#> [17,]  0.19683467 0.13774157  0.052817352 0.08732596 0.08664532 0.04634305
#> [18,]  0.12940000 0.09090000  0.080000000 0.05670000 0.01980000 0.01140000
#> [19,]  0.19999901 0.14617408  0.118118189 0.07659763 0.05390090 0.02463204
#> [20,]  0.17369666 0.16678899  0.069247403 0.11070730 0.11993198 0.06850611
#> [21,]  0.12112023 0.08904883  0.096077318 0.05721569 0.03421357 0.02062754
#> [22,]  0.20655349 0.23745988  0.206627031 0.19937090 0.17921599 0.12946628
#> [23,]  0.22220000 0.21110000  0.017600000 0.13480000 0.07440000 0.01300000
#> [24,]  0.22280000 0.15820000  0.143300000 0.16340000 0.11330000 0.05670000
#> [25,]  0.17423686 0.17536334  0.144386423 0.13668924 0.11993162 0.07004693
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.10413473 0.058391959 0.027609126 0.017148176 0.018815062 0.015534776
#>  [2,] 0.06350000 0.047500000 0.100400000 0.070900000 0.031700000 0.030900000
#>  [3,] 0.24950820 0.060481932 0.031816041 0.036153596 0.027456972 0.017612746
#>  [4,] 0.05307671 0.019516317 0.019131367 0.012282420 0.007643367 0.009541469
#>  [5,] 0.15882170 0.011809678 0.014106243 0.013139796 0.012024486 0.011086862
#>  [6,] 0.07047620 0.033493367 0.035063309 0.041037113 0.023014276 0.020322808
#>  [7,] 0.05925858 0.018590716 0.020665067 0.017236627 0.009635917 0.013000530
#>  [8,] 0.09791807 0.019024106 0.017371369 0.014994839 0.011972133 0.010730858
#>  [9,] 0.13158212 0.012914203 0.009315303 0.006701235 0.016161429 0.016763556
#> [10,] 0.08247315 0.019944687 0.011482282 0.029752125 0.015825926 0.015054183
#> [11,] 0.05541259 0.019405620 0.035987365 0.019229713 0.014375813 0.025478581
#> [12,] 0.03553106 0.021166565 0.012899971 0.009485680 0.009363335 0.013629466
#> [13,] 0.06687631 0.009921664 0.016196444 0.009088839 0.012286831 0.007670288
#> [14,] 0.06689339 0.010902367 0.012762550 0.007802367 0.002274408 0.012113733
#> [15,] 0.05810800 0.031823880 0.013925988 0.012984773 0.018189806 0.002135045
#> [16,] 0.04045169 0.020218125 0.016291215 0.012561999 0.017632795 0.020134767
#> [17,] 0.02844060 0.019492462 0.010210134 0.002764464 0.016193520 0.013229226
#> [18,] 0.17240000 0.015100000 0.008500000 0.017800000 0.007300000 0.007900000
#> [19,] 0.04866577 0.016196875 0.014400136 0.009748955 0.008293407 0.008034517
#> [20,] 0.11330452 0.006604818 0.023489287 0.010681319 0.009693513 0.008488102
#> [21,] 0.04474400 0.014365365 0.011040334 0.013499535 0.008674138 0.007693215
#> [22,] 0.04395380 0.034414313 0.015391235 0.011452015 0.006437887 0.012163594
#> [23,] 0.09740000 0.010600000 0.003300000 0.023200000 0.016600000 0.009500000
#> [24,] 0.04760000 0.013300000 0.017000000 0.003500000 0.005200000 0.008300000
#> [25,] 0.06579629 0.019715090 0.019022216 0.012944302 0.007723819 0.012243641
#>               V55         V56          V57         V58         V59         V6
#>  [1,] 0.013558749 0.012258673 0.0114770350 0.015891353 0.009877462 0.13394905
#>  [2,] 0.025200000 0.008700000 0.0177000000 0.021400000 0.022700000 0.13280000
#>  [3,] 0.023372265 0.003356291 0.0010707709 0.019897854 0.019327952 0.25798303
#>  [4,] 0.012196054 0.007320413 0.0071876078 0.007728972 0.007102176 0.10759413
#>  [5,] 0.003059620 0.003793150 0.0028879049 0.005363526 0.008319126 0.09176180
#>  [6,] 0.015248016 0.007745329 0.0176873543 0.011285345 0.015849763 0.07332353
#>  [7,] 0.008732784 0.011170676 0.0038961012 0.007257854 0.008139827 0.11770390
#>  [8,] 0.006287315 0.007935108 0.0066436469 0.005183021 0.004939351 0.12475234
#>  [9,] 0.001765424 0.017382802 0.0104156610 0.005946275 0.013651487 0.09657426
#> [10,] 0.024814698 0.012870854 0.0039979954 0.011868004 0.014439517 0.04311043
#> [11,] 0.016063692 0.009680106 0.0100237473 0.019695009 0.013146739 0.15756172
#> [12,] 0.008220074 0.006361822 0.0030140432 0.007864752 0.002909672 0.05104143
#> [13,] 0.003465486 0.007229084 0.0041137779 0.003203079 0.004232782 0.09219875
#> [14,] 0.005470612 0.007904734 0.0078388563 0.005574408 0.005623693 0.08457682
#> [15,] 0.005572089 0.002372807 0.0003151193 0.016472262 0.010670732 0.08173118
#> [16,] 0.012858197 0.010105027 0.0078145421 0.012178490 0.004140719 0.02670463
#> [17,] 0.007380322 0.007847179 0.0067152354 0.008015697 0.010159475 0.04433113
#> [18,] 0.003800000 0.011600000 0.0033000000 0.003900000 0.008100000 0.38230000
#> [19,] 0.007307798 0.006627459 0.0090393313 0.006382401 0.008216679 0.08488831
#> [20,] 0.005663715 0.007162046 0.0092935129 0.014410525 0.019656151 0.10635416
#> [21,] 0.007491474 0.006019982 0.0045162374 0.004519206 0.004429262 0.07148233
#> [22,] 0.005578347 0.005667949 0.0127086936 0.007418100 0.006832634 0.06461583
#> [23,] 0.018000000 0.024400000 0.0316000000 0.016400000 0.009500000 0.22800000
#> [24,] 0.007800000 0.007500000 0.0105000000 0.016000000 0.009500000 0.08160000
#> [25,] 0.010292309 0.004785023 0.0105655692 0.007466630 0.009623414 0.13449706
#>               V60         V7         V8         V9
#>  [1,] 0.005487781 0.14806288 0.15084535 0.27851761
#>  [2,] 0.010600000 0.09880000 0.17870000 0.11990000
#>  [3,] 0.013398476 0.21338767 0.10587867 0.15083038
#>  [4,] 0.005986519 0.09338567 0.14937356 0.23322461
#>  [5,] 0.002919723 0.06782888 0.12645239 0.33980230
#>  [6,] 0.005554660 0.03327157 0.07925255 0.19513844
#>  [7,] 0.007716976 0.13832047 0.14211811 0.21820219
#>  [8,] 0.004439688 0.13733082 0.15873578 0.20532129
#>  [9,] 0.009326324 0.04786164 0.11457397 0.06507436
#> [10,] 0.006733355 0.06896074 0.06165202 0.11432636
#> [11,] 0.007417421 0.19648211 0.23993882 0.26893902
#> [12,] 0.003877458 0.08890209 0.12131059 0.24618363
#> [13,] 0.003066819 0.13213658 0.09479246 0.12188580
#> [14,] 0.003844551 0.09795599 0.06101294 0.08804329
#> [15,] 0.008835889 0.08846720 0.10010217 0.13373325
#> [16,] 0.004686408 0.04717688 0.06063613 0.11741984
#> [17,] 0.008167578 0.02772915 0.05134005 0.02958656
#> [18,] 0.005300000 0.37290000 0.35830000 0.34290000
#> [19,] 0.005856276 0.10175570 0.09789176 0.11285887
#> [20,] 0.015979354 0.18058673 0.16705203 0.13744436
#> [21,] 0.003958302 0.08786776 0.08972532 0.12068282
#> [22,] 0.003496638 0.17152616 0.22912103 0.22998716
#> [23,] 0.007800000 0.24310000 0.37710000 0.55980000
#> [24,] 0.001100000 0.09930000 0.03150000 0.07360000
#> [25,] 0.008972397 0.10144646 0.19067924 0.28418752
#> 
#> $cl
#>  [1] M M M M M M M M M M M M R R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478