Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13       V14
#>  [1,] 0.02059241 0.18667849 0.10721500 0.02545178 0.10277816 0.2337423
#>  [2,] 0.06380287 0.32550543 0.48988263 0.48844261 0.47110317 0.4527638
#>  [3,] 0.00774328 0.12749892 0.18872294 0.21173678 0.20112771 0.1274000
#>  [4,] 0.02016157 0.17844477 0.19037300 0.23700408 0.26907247 0.2790926
#>  [5,] 0.02310126 0.22139968 0.26735710 0.29252278 0.25779867 0.1950544
#>  [6,] 0.02381326 0.25614778 0.28070237 0.26277312 0.41598824 0.5994150
#>  [7,] 0.05695033 0.32014761 0.24169117 0.18190459 0.39218155 0.5127176
#>  [8,] 0.01790875 0.31541848 0.29938092 0.30941598 0.24727276 0.1814024
#>  [9,] 0.02796264 0.19401579 0.24213490 0.26931850 0.29440715 0.3630277
#> [10,] 0.04637492 0.15243877 0.28741588 0.33896855 0.34308621 0.3291004
#> [11,] 0.03928838 0.18144978 0.22684210 0.25886785 0.27190137 0.2972744
#> [12,] 0.02191738 0.27928458 0.34107432 0.38206114 0.34856722 0.2620114
#> [13,] 0.02010255 0.37434458 0.34936577 0.36990597 0.40544149 0.4699708
#> [14,] 0.05916969 0.41750209 0.36905596 0.19026206 0.20768389 0.2444337
#> [15,] 0.01307376 0.09214241 0.11727145 0.08111339 0.05091436 0.0704537
#> [16,] 0.03631169 0.28415644 0.24884204 0.09956421 0.17244150 0.4770851
#> [17,] 0.01011448 0.10545626 0.09388865 0.10227067 0.15353759 0.1384337
#> [18,] 0.01069577 0.10030391 0.12686745 0.15699503 0.22132382 0.2661037
#> [19,] 0.02691575 0.08356162 0.08856981 0.17034047 0.17880473 0.1223504
#> [20,] 0.01954482 0.09744377 0.07942536 0.06081585 0.16667930 0.3349756
#> [21,] 0.01155194 0.16455487 0.13001996 0.15995666 0.31669290 0.2318663
#> [22,] 0.02972076 0.12153814 0.24868950 0.35689806 0.27785941 0.2619336
#> [23,] 0.01960335 0.13651095 0.21621279 0.16418015 0.14686466 0.1844867
#> [24,] 0.01744296 0.05183436 0.09684788 0.08619640 0.17998243 0.2673903
#> [25,] 0.01643356 0.11258096 0.14192995 0.14642094 0.13693635 0.1311856
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.42648109 0.63070323 0.73751457 0.7277052 0.8196396 0.019038798
#>  [2,] 0.55891394 0.76707928 0.93527016 0.9322910 0.8762978 0.114863311
#>  [3,] 0.03655301 0.14674786 0.27572280 0.3856897 0.4807134 0.020403851
#>  [4,] 0.24313623 0.20637988 0.17223733 0.2407231 0.3651734 0.033357983
#>  [5,] 0.16962338 0.17494883 0.18269116 0.2480203 0.2723381 0.019512705
#>  [6,] 0.64099883 0.68861630 0.66910868 0.7358733 0.8362843 0.032063817
#>  [7,] 0.22864606 0.38271031 0.54784821 0.5607090 0.7684649 0.060040984
#>  [8,] 0.18594929 0.22016550 0.21760512 0.2090942 0.1578227 0.023907406
#>  [9,] 0.43316698 0.50370133 0.57281586 0.6534772 0.7258473 0.022090469
#> [10,] 0.44415842 0.50186599 0.62872073 0.6320189 0.6619846 0.056911881
#> [11,] 0.30808431 0.33788243 0.32439064 0.2419254 0.3116881 0.054560772
#> [12,] 0.20316605 0.24038891 0.25757780 0.3374291 0.5299226 0.030515797
#> [13,] 0.48754053 0.63554270 0.81980495 0.8226490 0.8250644 0.094391060
#> [14,] 0.31431875 0.28580797 0.17642814 0.2260443 0.4663145 0.078342667
#> [15,] 0.11931647 0.09947281 0.09995060 0.1973173 0.2970931 0.017267484
#> [16,] 0.66050636 0.70375832 0.74053251 0.6000858 0.3933919 0.066928226
#> [17,] 0.11514156 0.10422792 0.09757061 0.1680376 0.2393279 0.009472976
#> [18,] 0.18326034 0.21119220 0.23665887 0.3225038 0.4387545 0.011163527
#> [19,] 0.14335612 0.19052065 0.20842801 0.2157706 0.2474163 0.043895597
#> [20,] 0.50174860 0.68955621 0.78226261 0.7883476 0.8243999 0.018631230
#> [21,] 0.12066976 0.27393989 0.36402905 0.3885957 0.6256358 0.046744064
#> [22,] 0.28934217 0.40996671 0.56443252 0.5012349 0.4810360 0.057057258
#> [23,] 0.18921045 0.20668558 0.24285951 0.1144011 0.3897325 0.005567309
#> [24,] 0.25683230 0.20052756 0.26670052 0.3237681 0.4149722 0.009374336
#> [25,] 0.10473419 0.16177138 0.25481104 0.3348441 0.3241387 0.014947294
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.9361833 0.9690646 1.0327940 0.8708871 0.5543932 0.3680028 0.3424254
#>  [2,] 0.8393521 0.8631811 0.5205298 0.1463017 0.1352371 0.2615599 0.1478003
#>  [3,] 0.5817109 0.6806269 0.7579740 0.7662867 0.7180460 0.7845039 0.9097181
#>  [4,] 0.3979403 0.4548680 0.5711910 0.6771504 0.7565741 0.8157188 0.8206528
#>  [5,] 0.3346434 0.4027385 0.4733620 0.5972518 0.7146508 0.8186532 0.8691205
#>  [6,] 0.9393836 0.9970595 0.9883606 0.9683538 0.8857233 0.7663591 0.7705341
#>  [7,] 0.8407423 0.5834332 0.6775658 0.8292752 0.6877951 0.7590959 0.8122762
#>  [8,] 0.1483625 0.1625513 0.2825271 0.4587605 0.5960505 0.7016427 0.8017867
#>  [9,] 0.7739196 0.8339669 0.8944934 0.9342306 0.9727862 0.9484060 0.9340702
#> [10,] 0.7739079 0.8172687 0.8711983 0.9607069 1.0939133 0.8962289 0.6493005
#> [11,] 0.3879769 0.4489455 0.4318615 0.4594999 0.5325470 0.5286611 0.5898109
#> [12,] 0.6713321 0.7941478 0.8181656 0.8164327 0.8927380 0.9001788 0.9354594
#> [13,] 0.8496354 0.8068862 0.8582992 1.0157053 1.0706202 0.9823331 0.9670190
#> [14,] 0.8239616 0.8031659 0.3996917 0.1659117 0.2119066 0.3178900 0.6235780
#> [15,] 0.3073381 0.3332397 0.5148701 0.7357272 0.9010635 0.9460105 0.7765529
#> [16,] 0.3199006 0.5408101 0.7745587 0.8610124 0.9291824 0.8219492 0.6805816
#> [17,] 0.2406432 0.2469112 0.3791494 0.5157334 0.6612589 0.7060326 0.7378209
#> [18,] 0.5163480 0.5998804 0.6512575 0.5664498 0.6354398 0.8727260 0.9408023
#> [19,] 0.2079047 0.2338419 0.1671771 0.2117827 0.2411958 0.2055477 0.3000413
#> [20,] 0.8411808 0.9204438 0.9119999 0.9105806 0.9187606 0.9023867 0.9536527
#> [21,] 0.7276159 0.6796280 0.4668229 0.4186434 0.4037519 0.6303625 0.7941321
#> [22,] 0.6480824 0.4512206 0.2933342 0.4413328 0.8910302 0.8693914 0.4647960
#> [23,] 0.5954303 0.4022087 0.4179016 1.0148252 0.8804704 0.4465835 0.5299995
#> [24,] 0.4652247 0.5158616 0.5823257 0.6592791 0.7273888 0.8630476 0.8752265
#> [25,] 0.4507388 0.5707327 0.7277388 0.8290108 0.8230877 0.8328889 0.8636834
#>             V27       V28         V29          V3       V30       V31       V32
#>  [1,] 0.2626386 0.2985821  0.46515499 0.002810843 0.5894103 0.3733546 0.3615257
#>  [2,] 0.1177207 0.1207025 -0.01047686 0.149400353 0.1702072 0.1106545 0.1547602
#>  [3,] 0.9682463 0.9630469  0.83941362 0.024523880 0.5846472 0.2369849 0.2232271
#>  [4,] 0.8655742 0.8765216  0.79231548 0.038323586 0.6430933 0.5996841 0.5551229
#>  [5,] 0.9275633 0.9779798  0.89227490 0.032453388 0.7088133 0.6457785 0.5016655
#>  [6,] 0.7231653 0.6401455  0.51045684 0.031330978 0.3672588 0.2745475 0.1898891
#>  [7,] 0.7303233 0.6556136  0.41790914 0.067908174 0.6733743 0.7427828 0.8229910
#>  [8,] 0.8931611 0.9185530  0.89175702 0.026872474 0.9203883 0.9032124 0.8409634
#>  [9,] 0.9084533 0.8379061  0.70891695 0.030554487 0.5705568 0.4362201 0.3171373
#> [10,] 0.4396876 0.2519613  0.27112541 0.086884075 0.3407164 0.3089078 0.1188285
#> [11,] 0.6931921 0.7948197  0.80982323 0.063746107 0.7722518 0.6935707 0.6177165
#> [12,] 0.9659840 0.9440664  0.82390781 0.039107824 0.5971702 0.4201993 0.3527366
#> [13,] 0.9653979 0.7339005  0.58899896 0.109518430 0.5139241 0.3084715 0.2322002
#> [14,] 0.6289542 0.7148269  0.69125054 0.054676301 0.7452739 0.4343447 0.4588216
#> [15,] 0.6958750 0.7085075  0.60265527 0.007977380 0.4002086 0.2399122 0.2698771
#> [16,] 0.6766264 0.5863071  0.43622716 0.098490899 0.4845871 0.4618440 0.2252750
#> [17,] 0.7147615 0.8250310  0.97740726 0.020908698 0.9507499 0.8510487 0.7971862
#> [18,] 1.0042732 0.9409659  0.79612799 0.006345776 0.6374345 0.5289733 0.3729997
#> [19,] 0.4079360 0.5956867  0.71407341 0.030833865 0.7455906 0.7518575 0.6144782
#> [20,] 0.9383648 0.7846783  0.51384459 0.024651541 0.4097680 0.3613895 0.3167132
#> [21,] 0.8009808 0.5137188  0.33944772 0.057468158 0.2664294 0.2466932 0.4190993
#> [22,] 0.3300808 0.6405469  0.76528450 0.065148020 0.4736272 0.5666492 0.7036937
#> [23,] 0.6344251 0.7428707  0.58024245 0.054074424 0.3883835 0.2552751 0.1920092
#> [24,] 0.9666023 0.9818329  0.80866252 0.016250824 0.6975467 0.6273621 0.5018245
#> [25,] 0.8256285 0.8003656  0.79446007 0.027840850 0.7213549 0.6398610 0.5855225
#>             V33        V34       V35        V36       V37       V38        V39
#>  [1,] 0.4637474 0.35614723 0.2214348 0.18571788 0.1485187 0.2001959 0.16458922
#>  [2,] 0.2662745 0.28053300 0.3905690 0.12343712 0.1016274 0.2399942 0.17885872
#>  [3,] 0.3129789 0.48935257 0.7453470 0.87875530 0.8137638 0.6583137 0.50696782
#>  [4,] 0.5057701 0.43953801 0.2668518 0.16461312 0.1725878 0.1784613 0.31095746
#>  [5,] 0.4002946 0.26454792 0.1397083 0.08738378 0.1722076 0.1804746 0.20838613
#>  [6,] 0.0923912 0.08165225 0.1548208 0.26082097 0.2376912 0.2212259 0.27471958
#>  [7,] 0.7131032 0.59519506 0.3738830 0.13257310 0.1160014 0.3876225 0.59194351
#>  [8,] 0.7185849 0.65572001 0.6721778 0.68504725 0.6274860 0.5383546 0.45645725
#>  [9,] 0.2207367 0.13014563 0.1517107 0.16778886 0.2008507 0.2292135 0.23047441
#> [10,] 0.3332859 0.45531878 0.4205815 0.25888164 0.2941775 0.3992059 0.39382044
#> [11,] 0.4958640 0.39681064 0.3049101 0.21173154 0.2260570 0.3093875 0.31658318
#> [12,] 0.2823124 0.22444160 0.1585606 0.14728311 0.1495884 0.2077155 0.25715406
#> [13,] 0.1304911 0.30186420 0.3856529 0.25820581 0.2476521 0.2732516 0.21506949
#> [14,] 0.7078105 0.74707621 0.7499567 0.67077181 0.6357566 0.6656985 0.56033262
#> [15,] 0.2798438 0.28600088 0.1952555 0.25217749 0.3521079 0.2972353 0.29601526
#> [16,] 0.2017994 0.10613907 0.4114896 0.67995757 0.7058479 0.3570212 0.16064196
#> [17,] 0.8079843 0.69315674 0.5678915 0.47599068 0.3639179 0.2845268 0.25317106
#> [18,] 0.2695553 0.16398831 0.1090306 0.24161003 0.2727473 0.3359281 0.32919378
#> [19,] 0.6497576 0.77064322 0.8397937 0.85806987 0.7327325 0.6667824 0.65287620
#> [20,] 0.2364496 0.23887008 0.2130010 0.13665295 0.1932067 0.1544472 0.06651232
#> [21,] 0.4857364 0.37349833 0.3460601 0.48503365 0.6090139 0.4208990 0.55718482
#> [22,] 0.6144602 0.87411273 0.7709597 0.71044666 0.7532916 0.4990641 0.36382885
#> [23,] 0.1159060 0.31142483 0.2692514 0.33357278 0.2664504 0.1956042 0.23453714
#> [24,] 0.3858018 0.19290603 0.1193624 0.24645840 0.2405134 0.3023563 0.32275766
#> [25,] 0.6044053 0.62718864 0.6330542 0.57503137 0.4584216 0.3280251 0.21278823
#>                V4        V40        V41         V42         V43        V44
#>  [1,] 0.040162445 0.10332471 0.15535946  0.18016001 0.115079440 0.16686913
#>  [2,] 0.147628796 0.04550467 0.16380480  0.20043006 0.170093252 0.11990111
#>  [3,] 0.024020661 0.34383830 0.26818701  0.23847259 0.328105360 0.39320271
#>  [4,] 0.016517597 0.39330058 0.34560652  0.26306073 0.194194084 0.10997683
#>  [5,] 0.039109088 0.16586607 0.18955145  0.23524845 0.181180074 0.12563287
#>  [6,] 0.046613819 0.26302003 0.25050286  0.25047841 0.228081104 0.22941558
#>  [7,] 0.102621118 0.41757324 0.46624251  0.50285005 0.423220719 0.36724841
#>  [8,] 0.059843971 0.44194858 0.47609891  0.44126831 0.419570448 0.51031101
#>  [9,] 0.062330366 0.25297730 0.25982535  0.25789937 0.223161544 0.16937636
#> [10,] 0.032030745 0.35982974 0.27709553  0.34954116 0.280502965 0.22270380
#> [11,] 0.064945540 0.29990733 0.46225085  0.50676704 0.410756997 0.35728333
#> [12,] 0.036645789 0.22894261 0.12070471  0.16797387 0.246916563 0.26346669
#> [13,] 0.047652973 0.20168774 0.33559650  0.29051851 0.004678902 0.12859828
#> [14,] 0.079166003 0.38545704 0.33311585  0.33981866 0.298057030 0.26272411
#> [15,] 0.002672947 0.29254787 0.25811633  0.22299023 0.169694127 0.19076486
#> [16,] 0.084381680 0.37168376 0.44356092  0.37633576 0.246978084 0.19199373
#> [17,] 0.030230650 0.26191274 0.21271747  0.16876875 0.200579207 0.21660035
#> [18,] 0.022524787 0.20304597 0.10068639  0.18427827 0.147433898 0.12829435
#> [19,] 0.015694079 0.62594150 0.53432763  0.49480194 0.380835205 0.28272530
#> [20,] 0.041091769 0.08158389 0.07633312  0.04077771 0.063956671 0.09319966
#> [21,] 0.040505289 0.48982459 0.29452913  0.27171760 0.244951773 0.23062295
#> [22,] 0.093078971 0.35484666 0.30040778  0.26822753 0.070273508 0.18682460
#> [23,] 0.053204291 0.10983797 0.02676822 -0.00189474 0.055807612 0.09440376
#> [24,] 0.013492019 0.20445355 0.12743182  0.25853455 0.206156061 0.14208156
#> [25,] 0.039552289 0.25931118 0.24969132  0.22678136 0.226901665 0.15359817
#>              V45        V46        V47         V48        V49         V5
#>  [1,] 0.17230935 0.05322709 0.05477424 0.066819095 0.04154446 0.06868953
#>  [2,] 0.06953734 0.03142161 0.08882726 0.083204148 0.03107909 0.16886967
#>  [3,] 0.34270873 0.25365493 0.13534643 0.072160742 0.04539311 0.04251307
#>  [4,] 0.08410854 0.10069721 0.10233444 0.093729976 0.04919568 0.03148437
#>  [5,] 0.09126863 0.12146456 0.14099827 0.107644508 0.04968637 0.06534025
#>  [6,] 0.14634758 0.11261607 0.10746656 0.106065069 0.07872656 0.12590990
#>  [7,] 0.49923893 0.51384638 0.29503492 0.224201977 0.12682355 0.21331948
#>  [8,] 0.49904786 0.35371338 0.24423837 0.187375558 0.10087602 0.07331836
#>  [9,] 0.12643352 0.10165811 0.10484326 0.074553492 0.04561890 0.11305505
#> [10,] 0.23915638 0.13580431 0.09832990 0.083068917 0.04548267 0.06496333
#> [11,] 0.42768883 0.38764426 0.26909990 0.179962459 0.10451598 0.07699185
#> [12,] 0.25347053 0.21114483 0.15727051 0.107545533 0.05920132 0.06822052
#> [13,] 0.17903105 0.15758986 0.09216493 0.065054746 0.01561265 0.12820087
#> [14,] 0.29517086 0.23575750 0.11316846 0.091424008 0.04577545 0.11880079
#> [15,] 0.17591096 0.14613299 0.06403859 0.058638100 0.03685999 0.01451056
#> [16,] 0.21399535 0.18579700 0.18236559 0.165895252 0.11144720 0.06106959
#> [17,] 0.19127196 0.11427276 0.05942436 0.039770971 0.01930323 0.02701221
#> [18,] 0.09882287 0.06260363 0.09571870 0.081191981 0.03047117 0.01730434
#> [19,] 0.21787075 0.15609621 0.14958194 0.118276226 0.06581953 0.04210117
#> [20,] 0.07772399 0.02778230 0.05607983 0.042656533 0.02439268 0.06491908
#> [21,] 0.11262199 0.07736264 0.06054905 0.041791165 0.05131371 0.03720960
#> [22,] 0.15606963 0.16319458 0.11464620 0.067167392 0.04335392 0.15663350
#> [23,] 0.05876261 0.02359014 0.02503065 0.005913224 0.01110231 0.06584280
#> [24,] 0.12373680 0.08874415 0.11422993 0.085344245 0.04326377 0.04626226
#> [25,] 0.10088112 0.07514416 0.05002365 0.044949699 0.01876050 0.05848883
#>                V50           V51         V52         V53           V54
#>  [1,] 0.0143152722  1.561320e-02 0.011058252 0.001026915  0.0040740947
#>  [2,] 0.0201256685  2.656861e-02 0.023895974 0.027904270  0.0137275907
#>  [3,] 0.0156835311  7.215522e-03 0.009501956 0.006459225  0.0058031694
#>  [4,] 0.0105694558  1.266837e-02 0.013437040 0.008998028  0.0038496738
#>  [5,] 0.0157230909  1.552507e-02 0.010125427 0.008136637  0.0064556580
#>  [6,] 0.0204808120  2.406212e-02 0.019426553 0.015901469  0.0187681770
#>  [7,] 0.0742376038  2.071208e-02 0.018353290 0.029286951  0.0196584306
#>  [8,] 0.0285107428  1.905894e-02 0.017024961 0.012182032  0.0108679052
#>  [9,] 0.0207815026  1.751201e-02 0.015188729 0.015129279  0.0132719114
#> [10,] 0.0008150885  6.652394e-04 0.005288622 0.015212858  0.0114588301
#> [11,] 0.0326998795  3.227679e-02 0.024852841 0.013987973  0.0140121989
#> [12,] 0.0171024174  1.989698e-02 0.012459011 0.006412795  0.0070290547
#> [13,] 0.0154807353  1.677679e-02 0.019741664 0.022036039  0.0291148280
#> [14,] 0.0192559752  1.950235e-02 0.018571413 0.012598412  0.0150570261
#> [15,] 0.0058512713 -2.424318e-06 0.006588831 0.007751095  0.0105482814
#> [16,] 0.0280521636  1.772968e-02 0.012266593 0.008733466  0.0122383511
#> [17,] 0.0122821726  9.770879e-03 0.008272772 0.010853458  0.0068106698
#> [18,] 0.0118308131  6.062355e-03 0.006070228 0.008259410  0.0075574813
#> [19,] 0.0213798377  1.332371e-02 0.009437956 0.016920507  0.0111777073
#> [20,] 0.0145309865  1.173855e-02 0.009765141 0.007620898  0.0080229102
#> [21,] 0.0195272664  9.326932e-03 0.008958966 0.004009697  0.0206738424
#> [22,] 0.0086581029  1.572405e-02 0.006329510 0.014469332  0.0066547115
#> [23,] 0.0108098822  3.256768e-02 0.043240969 0.015501125 -0.0008561107
#> [24,] 0.0240742469  3.483687e-03 0.004880244 0.005116682  0.0070378862
#> [25,] 0.0115692106  6.177839e-03 0.007146399 0.006751950  0.0053545003
#>                V55         V56         V57         V58          V59         V6
#>  [1,] 0.0014609696 0.004358086 0.001223398 0.004053832 0.0039617014 0.13654955
#>  [2,] 0.0209224328 0.016235367 0.016735332 0.009758790 0.0178962563 0.09142657
#>  [3,] 0.0032333338 0.003607715 0.003059817 0.001742912 0.0033922992 0.06014088
#>  [4,] 0.0052709068 0.006875452 0.003958273 0.003999564 0.0040948694 0.06120737
#>  [5,] 0.0050715457 0.005809388 0.003924141 0.004257954 0.0031878382 0.09712351
#>  [6,] 0.0159723814 0.005375261 0.007759217 0.016381848 0.0109238506 0.17565577
#>  [7,] 0.0365547151 0.031412063 0.029520829 0.036481085 0.0209901322 0.12686141
#>  [8,] 0.0084783708 0.007953176 0.007398641 0.005132498 0.0052756593 0.10870395
#>  [9,] 0.0053496242 0.010442440 0.008121392 0.009355957 0.0076844934 0.13393234
#> [10,] 0.0079080936 0.013023417 0.008688049 0.008759936 0.0140219988 0.09250135
#> [11,] 0.0134647579 0.010276758 0.011090750 0.011417063 0.0118385537 0.09430049
#> [12,] 0.0071734809 0.004776351 0.004704421 0.004181985 0.0036564295 0.08856502
#> [13,] 0.0145157525 0.013319020 0.012625672 0.010042686 0.0007839582 0.17751394
#> [14,] 0.0105513789 0.006720616 0.005412918 0.009865384 0.0107892254 0.10087471
#> [15,] 0.0004061982 0.001460701 0.002518897 0.005389326 0.0036225001 0.01308400
#> [16,] 0.0072153818 0.006319989 0.012511810 0.008461162 0.0072114537 0.08561088
#> [17,] 0.0039966167 0.005628292 0.003543709 0.002260164 0.0081010126 0.03086971
#> [18,] 0.0038022046 0.006437460 0.003098484 0.002964413 0.0030284984 0.02828174
#> [19,] 0.0115322200 0.008727172 0.005533084 0.011792026 0.0113964258 0.06790580
#> [20,] 0.0042249160 0.009385618 0.007834087 0.003959822 0.0052062332 0.08265045
#> [21,] 0.0221529130 0.006890066 0.014438150 0.020914848 0.0132964205 0.01831634
#> [22,] 0.0178916909 0.007462626 0.013321756 0.005689375 0.0168502831 0.23476290
#> [23,] 0.0097879194 0.009543703 0.008288606 0.019026786 0.0115307193 0.15995002
#> [24,] 0.0062454421 0.004998471 0.006521235 0.005902384 0.0011531575 0.03877838
#> [25,] 0.0053359861 0.006586460 0.009125378 0.006521567 0.0045836816 0.08703640
#>               V60         V7         V8         V9
#>  [1,] 0.001929870 0.14814608 0.12996939 0.19286713
#>  [2,] 0.014112802 0.07206868 0.02682370 0.15209360
#>  [3,] 0.003936744 0.07084997 0.02087712 0.08921007
#>  [4,] 0.004951911 0.07704216 0.06582349 0.13874909
#>  [5,] 0.003189746 0.11220082 0.12338158 0.16615711
#>  [6,] 0.004847760 0.18012762 0.17889457 0.21763876
#>  [7,] 0.009272967 0.16852486 0.10747836 0.24108828
#>  [8,] 0.004479709 0.13555139 0.13211722 0.28590021
#>  [9,] 0.005477163 0.14310699 0.15034613 0.17318115
#> [10,] 0.008093040 0.13760312 0.06558130 0.15237941
#> [11,] 0.007579459 0.10058198 0.14175973 0.18679990
#> [12,] 0.004891500 0.11324791 0.14496360 0.20899835
#> [13,] 0.002939136 0.16876605 0.22934698 0.34787786
#> [14,] 0.008846713 0.15801653 0.25788135 0.44993304
#> [15,] 0.004569072 0.06349210 0.09680956 0.06649560
#> [16,] 0.005517978 0.14589724 0.21572635 0.24324623
#> [17,] 0.004561340 0.05190623 0.05212365 0.07229171
#> [18,] 0.003288910 0.06553740 0.07147895 0.06865674
#> [19,] 0.005799199 0.07767404 0.09633379 0.07002586
#> [20,] 0.003692475 0.12077925 0.09548978 0.11046447
#> [21,] 0.004841595 0.03238802 0.02161680 0.07657649
#> [22,] 0.006788390 0.22279744 0.19490885 0.16829911
#> [23,] 0.006910395 0.14229685 0.21425251 0.14709370
#> [24,] 0.002253449 0.08902091 0.04799295 0.09306839
#> [25,] 0.006720231 0.08871689 0.10987033 0.09463443
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913