Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.025969354 0.37419087 0.37606046 0.37003834 0.49306731 0.45729196
#>  [2,] 0.041506421 0.30032023 0.33017627 0.29810614 0.23323962 0.10709657
#>  [3,] 0.031083669 0.08621829 0.15647228 0.22642102 0.30162667 0.37511398
#>  [4,] 0.030170079 0.29657058 0.36275685 0.36891860 0.44039085 0.54377106
#>  [5,] 0.045745926 0.23726467 0.24114271 0.18027649 0.21774921 0.32300630
#>  [6,] 0.017760668 0.27511275 0.31008983 0.35456310 0.29837194 0.21189619
#>  [7,] 0.008056676 0.24963028 0.26596102 0.28678735 0.20302562 0.05775604
#>  [8,] 0.034000000 0.41170000 0.52450000 0.53410000 0.55540000 0.39150000
#>  [9,] 0.030970741 0.20630805 0.30148853 0.33020020 0.29999015 0.23173813
#> [10,] 0.035212310 0.43953560 0.52968658 0.55750474 0.52330285 0.39021067
#> [11,] 0.016712715 0.58929527 0.64782791 0.47235951 0.38692930 0.30570958
#> [12,] 0.071412870 0.38275256 0.34280214 0.18719310 0.18371241 0.33124150
#> [13,] 0.040387958 0.22879660 0.23775055 0.25977984 0.31116815 0.29868978
#> [14,] 0.055389224 0.34768190 0.48772585 0.47985464 0.50126853 0.49716647
#> [15,] 0.017072596 0.10130211 0.10294717 0.07890338 0.07247352 0.16755814
#> [16,] 0.030380006 0.26310348 0.21736079 0.22725196 0.25669862 0.12361549
#> [17,] 0.014817987 0.11884063 0.16567340 0.21185487 0.29297177 0.29238225
#> [18,] 0.027049407 0.14609591 0.16407859 0.13193171 0.13244391 0.13132974
#> [19,] 0.026092306 0.24824218 0.22667940 0.09487859 0.18643182 0.44385014
#> [20,] 0.020006495 0.06810171 0.06784146 0.22589652 0.19234789 0.10956025
#> [21,] 0.011437044 0.09418064 0.08410995 0.08262141 0.13978793 0.14599309
#> [22,] 0.012600000 0.42840000 0.30150000 0.12070000 0.32990000 0.57070000
#> [23,] 0.015151334 0.19012297 0.14997235 0.08977495 0.09747890 0.11746493
#> [24,] 0.011096784 0.11066056 0.06864909 0.08495943 0.08451131 0.07607215
#> [25,] 0.021991115 0.01200727 0.04906886 0.17261219 0.12795987 0.18680575
#>             V15        V16       V17       V18       V19           V2       V20
#>  [1,] 0.3165639 0.37306225 0.3939655 0.3455942 0.4985639  0.042141590 0.7373688
#>  [2,] 0.1249642 0.17863714 0.1920705 0.2510016 0.2131442  0.039239135 0.1686478
#>  [3,] 0.4374049 0.55608179 0.6401626 0.7122006 0.8302048  0.037209327 0.9001409
#>  [4,] 0.5729274 0.57744308 0.5847813 0.6743342 0.7740076  0.027229638 0.8425863
#>  [5,] 0.4623689 0.62148593 0.7330652 0.6827789 0.7550267  0.047290217 0.8177087
#>  [6,] 0.1570145 0.21821003 0.2560360 0.4021886 0.5985003  0.027615993 0.6853273
#>  [7,] 0.1264629 0.30975961 0.4528942 0.5625553 0.6466320  0.001266942 0.6489349
#>  [8,] 0.2950000 0.30750000 0.3021000 0.2719000 0.5443000  0.062500000 0.7932000
#>  [9,] 0.1937162 0.18000790 0.2455012 0.2520755 0.3183425  0.045261720 0.3217222
#> [10,] 0.2453492 0.26029687 0.3039246 0.3235314 0.5192846  0.034028024 0.6801933
#> [11,] 0.1726077 0.07954879 0.1783389 0.4005745 0.6245624  0.021659084 0.7382162
#> [12,] 0.4836764 0.50578086 0.2606564 0.2088167 0.3180692  0.086802423 0.6490340
#> [13,] 0.2730100 0.28350218 0.3102976 0.2095247 0.2320648  0.056698748 0.3101048
#> [14,] 0.6194499 0.78393614 0.9144113 0.9375407 0.8558401  0.096589963 0.7910735
#> [15,] 0.1234892 0.10480975 0.1816871 0.2515575 0.2360356  0.013369682 0.3773826
#> [16,] 0.3461898 0.32408419 0.2433629 0.2318733 0.2424222 -0.001141786 0.3712805
#> [17,] 0.3340486 0.37202050 0.3213630 0.3245849 0.4738613  0.013930367 0.6075902
#> [18,] 0.2179984 0.55253174 0.7037343 0.5740603 0.3152576  0.007796029 0.3328354
#> [19,] 0.5713092 0.68885475 0.7719087 0.7078242 0.5246106  0.059601202 0.4019177
#> [20,] 0.1407003 0.17754897 0.1988717 0.2375854 0.1828280  0.025855255 0.1259611
#> [21,] 0.1579065 0.18268409 0.1966178 0.2490230 0.3357000  0.012723066 0.3490117
#> [22,] 0.6962000 0.97510000 1.0000000 0.9293000 0.6210000  0.051900000 0.4586000
#> [23,] 0.1568654 0.20891348 0.2432091 0.2421056 0.3571033  0.002045780 0.5587919
#> [24,] 0.0467675 0.11979232 0.2407835 0.2706709 0.3409641  0.019918861 0.4624658
#> [25,] 0.2137978 0.34624779 0.7596750 0.6278058 0.2499635  0.031594191 0.1757197
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.7153102 0.65367628 0.6166582 0.6001844 0.66452136 0.8538646 0.9248223
#>  [2,] 0.3290605 0.61665716 0.7808560 0.8468501 0.93150370 0.9973593 0.9833684
#>  [3,] 0.9275440 0.95808626 0.9236975 0.8500321 0.76124243 0.6752481 0.6707482
#>  [4,] 0.8670965 0.92699825 0.9768331 0.9812862 0.89125015 0.8451194 0.8592521
#>  [5,] 0.9141085 0.95375186 0.8445445 0.5523371 0.29422744 0.3987092 0.5295443
#>  [6,] 0.7830709 0.78140719 0.8231042 0.9281326 0.93201388 0.9507713 0.9658484
#>  [7,] 0.7075248 0.82278179 0.9373086 1.0197799 1.01105436 0.8346459 0.5652220
#>  [8,] 0.8751000 0.86670000 0.7107000 0.6911000 0.72870000 0.8792000 1.0000000
#>  [9,] 0.4505662 0.57364082 0.6383960 0.7035981 0.72681864 0.7610322 0.8850002
#> [10,] 0.8099866 0.87128761 0.8070085 0.8210617 0.81216637 0.8191214 0.8892040
#> [11,] 0.8098738 0.87362825 0.8815818 0.9378462 0.88528832 0.8074451 0.7598650
#> [12,] 0.6883757 0.29029202 0.1531813 0.2888968 0.23380435 0.3427843 0.3119448
#> [13,] 0.3307669 0.40867705 0.5101300 0.5349916 0.58446789 0.6357113 0.7062455
#> [14,] 0.7424908 0.44535901 0.1441363 0.1085210 0.25885203 0.2586131 0.2088401
#> [15,] 0.4900832 0.64190805 0.8144241 0.8475648 0.89735151 0.9241930 0.8731956
#> [16,] 0.2439839 0.66901578 0.6561578 0.3431509 0.23735030 0.5709605 0.8875154
#> [17,] 0.5636913 0.48090005 0.5948936 0.6916229 0.68635622 0.7812385 0.8482124
#> [18,] 0.4586660 0.57044216 0.5821443 0.6164807 0.81567445 0.9854408 0.9818142
#> [19,] 0.5572108 0.75816101 0.8340093 0.8482692 0.72930437 0.6582928 0.6179162
#> [20,] 0.1965723 0.09907815 0.1565184 0.1768792 0.09148492 0.1968237 0.3770920
#> [21,] 0.3710341 0.41872551 0.4602240 0.5707000 0.60479010 0.6269190 0.6859207
#> [22,] 0.5001000 0.50320000 0.7082000 0.8420000 0.81090000 0.7690000 0.8105000
#> [23,] 0.6652704 0.86452919 1.0141037 0.9093095 0.84404708 0.9147792 0.9230189
#> [24,] 0.5946204 0.77316421 0.8932558 0.8606535 0.79899685 0.8698451 0.9757782
#> [25,] 0.3817373 0.63609126 0.7253323 0.7492638 0.85008595 0.9123953 0.8715080
#>             V28        V29           V3       V30       V31       V32       V33
#>  [1,] 0.8689810 0.71576676 0.0598589215 0.6235941 0.5812951 0.7142242 0.6102721
#>  [2,] 0.9936556 0.85701051 0.0208799603 0.6947619 0.7569440 0.6927618 0.4892433
#>  [3,] 0.6041516 0.48930126 0.0546945724 0.3836996 0.2505566 0.1941251 0.2567531
#>  [4,] 0.8406535 0.73970680 0.0384918201 0.5241007 0.3270200 0.2425009 0.1291029
#>  [5,] 0.4760772 0.38326915 0.0354418192 0.4896271 0.2697997 0.4272070 0.4759330
#>  [6,] 0.9493042 0.84667532 0.0352987132 0.6241052 0.3863471 0.2854267 0.2272917
#>  [7,] 0.2836656 0.00371936 0.0004561285 0.2815345 0.4977034 0.5678753 0.6182005
#>  [8,] 0.9816000 0.89840000 0.0381000000 0.6048000 0.4934000 0.5371000 0.4586000
#>  [9,] 0.9482038 0.86503463 0.0571217466 0.7177704 0.6637689 0.5411199 0.3700160
#> [10,] 0.8773112 0.69436463 0.0429493770 0.4672527 0.3691994 0.3866472 0.3753872
#> [11,] 0.7567956 0.58840207 0.0365533784 0.4150604 0.3906798 0.4351458 0.6643210
#> [12,] 0.4682342 0.44806620 0.0826192772 0.5900987 0.3370085 0.3707963 0.5528844
#> [13,] 0.8354376 0.83900638 0.0713855793 0.8349254 0.8816753 0.8322986 0.6832159
#> [14,] 0.1759790 0.03393970 0.1334084133 0.2397394 0.2110701 0.2322402 0.2917890
#> [15,] 0.7928360 0.75382366 0.0243182911 0.6731975 0.6296756 0.5304837 0.5385548
#> [16,] 0.7074617 0.73312150 0.0067232722 0.9006394 0.6153643 0.3380796 0.6949502
#> [17,] 0.8316954 0.69881196 0.0132776734 0.5744546 0.5067634 0.3683936 0.2950050
#> [18,] 0.9133433 0.73445670 0.0128448790 0.7669795 0.8457313 0.7338084 0.6663013
#> [19,] 0.4800874 0.34149155 0.0907683311 0.3937050 0.3803978 0.2744781 0.2373369
#> [20,] 0.5886059 0.65144196 0.0271334021 0.6249785 0.6944511 0.6309662 0.6731829
#> [21,] 0.8394359 0.95982031 0.0193591288 0.9090000 0.8059374 0.7384847 0.7435313
#> [22,] 0.6203000 0.23560000 0.0621000000 0.2595000 0.6299000 0.6762000 0.2903000
#> [23,] 0.8914955 0.80902961 0.0023586926 0.7028229 0.6341219 0.5316145 0.4783429
#> [24,] 0.9583683 0.94116376 0.0223638029 0.8394778 0.7744185 0.7422751 0.7013413
#> [25,] 0.7672748 0.86515467 0.0446066363 0.9286245 0.8818067 0.7339155 0.6755831
#>              V34        V35        V36        V37        V38        V39
#>  [1,] 0.38162259 0.13654474 0.27758222 0.31552708 0.40817137 0.58592152
#>  [2,] 0.19867444 0.11642310 0.12148840 0.12870999 0.15278177 0.18895589
#>  [3,] 0.28884003 0.38910788 0.44900894 0.41038229 0.34555797 0.35578163
#>  [4,] 0.01211317 0.08916271 0.17683632 0.21083747 0.28011936 0.28029862
#>  [5,] 0.26982889 0.15813477 0.22701101 0.30179812 0.38863375 0.30138678
#>  [6,] 0.19031879 0.16498319 0.07800322 0.05977953 0.08705232 0.11860637
#>  [7,] 0.66809554 0.55963218 0.37279405 0.18208210 0.15138892 0.21003694
#>  [8,] 0.29080000 0.07740000 0.22490000 0.16020000 0.39580000 0.61170000
#>  [9,] 0.29691060 0.18368249 0.15891234 0.27469296 0.17495108 0.15157973
#> [10,] 0.30764835 0.12892866 0.22769805 0.27108550 0.29219022 0.47052907
#> [11,] 0.72246617 0.59166730 0.46545109 0.38534618 0.31835687 0.40024407
#> [12,] 0.63472577 0.66750882 0.67595754 0.62335979 0.70019959 0.62160400
#> [13,] 0.56268729 0.53489431 0.41186540 0.30862991 0.34182383 0.30305295
#> [14,] 0.28787001 0.36996120 0.12926424 0.14060626 0.27548030 0.20436151
#> [15,] 0.61182230 0.58534560 0.47482876 0.34374434 0.25962236 0.15029958
#> [16,] 0.91098965 0.78728103 0.74171783 0.53619354 0.40219834 0.01484902
#> [17,] 0.19692565 0.25335972 0.38613365 0.35964478 0.21758318 0.23143896
#> [18,] 0.60703601 0.62332771 0.59449499 0.57039659 0.52698736 0.29155597
#> [19,] 0.11902037 0.35955080 0.61581709 0.63139117 0.35831325 0.19769210
#> [20,] 0.82807428 0.90504211 0.91518870 0.81986737 0.73155871 0.70605280
#> [21,] 0.66038846 0.58051041 0.51775494 0.42140755 0.29840726 0.19603068
#> [22,] 0.43930000 0.85290000 0.71800000 0.48010000 0.58560000 0.49930000
#> [23,] 0.51995579 0.52868249 0.46544355 0.45875222 0.33240319 0.18552051
#> [24,] 0.71750254 0.71143754 0.62330375 0.54667728 0.45871949 0.32687976
#> [25,] 0.77899632 0.93591354 1.04162788 0.93873609 0.75655002 0.60108033
#>                V4        V40        V41        V42        V43         V44
#>  [1,] 0.060714747 0.41506397 0.32769978 0.44917610 0.43815371 0.466452340
#>  [2,] 0.041022869 0.05941844 0.14644272 0.05940797 0.10087232 0.144938953
#>  [3,] 0.108361780 0.39712991 0.42250545 0.43079101 0.35505384 0.248460947
#>  [4,] 0.059586314 0.26107141 0.29375330 0.24760262 0.21156604 0.206894510
#>  [5,] 0.040121557 0.19970751 0.09489459 0.23775518 0.27672391 0.183500851
#>  [6,] 0.031780012 0.13523036 0.08477838 0.11883022 0.18734887 0.211902900
#>  [7,] 0.006682583 0.34359742 0.28112964 0.22848042 0.24408512 0.203420611
#>  [8,] 0.025700000 0.51960000 0.23210000 0.43700000 0.37970000 0.432200000
#>  [9,] 0.079256827 0.27329909 0.33536172 0.33559677 0.24043973 0.136856908
#> [10,] 0.032235244 0.42252782 0.16256530 0.23885921 0.37967039 0.443489007
#> [11,] 0.050197250 0.44925265 0.34766156 0.28400812 0.24744371 0.202820289
#> [12,] 0.109357581 0.48847412 0.37149452 0.28014447 0.22003108 0.230005336
#> [13,] 0.075850696 0.30174993 0.54274859 0.56759487 0.44227141 0.469380080
#> [14,] 0.124410237 0.09756798 0.20107063 0.21807308 0.14370376 0.130932176
#> [15,] 0.025690143 0.21740044 0.22154896 0.18009441 0.15785469 0.064688225
#> [16,] 0.034023334 0.62828226 0.77920463 0.62893476 0.33749678 0.145054027
#> [17,] 0.002563142 0.15305344 0.18680879 0.22306204 0.18886149 0.108627565
#> [18,] 0.026140762 0.14280466 0.19073982 0.19486557 0.21249027 0.166358255
#> [19,] 0.085858385 0.37792463 0.44920122 0.35621162 0.21174212 0.176912928
#> [20,] 0.032424929 0.69907624 0.62160332 0.53943450 0.38671060 0.224742806
#> [21,] 0.017787556 0.15545736 0.13143527 0.17223454 0.21884914 0.233276100
#> [22,] 0.051800000 0.28660000 0.06010000 0.11670000 0.27370000 0.281200000
#> [23,] 0.031726227 0.09869476 0.09038890 0.01530323 0.06498715 0.007055594
#> [24,] 0.039734552 0.24543148 0.17736959 0.09570939 0.05331531 0.059562935
#> [25,] 0.023236149 0.38210103 0.27266648 0.20501175 0.26387678 0.174280630
#>                V45        V46        V47         V48         V49           V5
#>  [1,]  0.518485321 0.28913898 0.19011926 0.209627493 0.108297903  0.106918402
#>  [2,]  0.101820167 0.07919735 0.10889903 0.108054157 0.065930996  0.052269695
#>  [3,]  0.184219286 0.11088449 0.10852707 0.074888437 0.041519843  0.150836793
#>  [4,]  0.123919930 0.07391330 0.09325542 0.087367235 0.061976643  0.117788913
#>  [5,]  0.063028896 0.05087961 0.06535862 0.065262531 0.050482842  0.062206237
#>  [6,]  0.210788929 0.22309990 0.18307089 0.115844343 0.055962717  0.076752416
#>  [7,]  0.153214617 0.14162645 0.12489240 0.093474600 0.046453208  0.044059070
#>  [8,]  0.489200000 0.19010000 0.09400000 0.136400000 0.090600000  0.044100000
#>  [9,]  0.170123914 0.19520268 0.21761943 0.195628464 0.091361482  0.076202776
#> [10,]  0.409308785 0.19289136 0.09537056 0.110706731 0.071342864  0.079621097
#> [11,]  0.176473403 0.12747908 0.12148386 0.131221524 0.064809081  0.034216155
#> [12,]  0.364287455 0.29144313 0.15650737 0.151750730 0.081903237  0.084214197
#> [13,]  0.575710224 0.51108635 0.32823985 0.189217874 0.097077659  0.070716299
#> [14,]  0.065939157 0.03264436 0.08264087 0.071353898 0.023386065  0.117574489
#> [15,]  0.019804001 0.01962759 0.03453231 0.045320824 0.014570069  0.046078643
#> [16,]  0.203888253 0.34499085 0.24156217 0.047351513 0.010210923  0.066712547
#> [17,]  0.076579352 0.07912951 0.04859502 0.033134484 0.020189817  0.014705678
#> [18,]  0.124587003 0.13074127 0.06935823 0.042728703 0.030198476  0.039668395
#> [19,]  0.176532120 0.11930044 0.13725510 0.136442414 0.084410234  0.099161879
#> [20,]  0.162286994 0.13209975 0.12987207 0.108784906 0.062545287  0.044039417
#> [21,]  0.187559184 0.12445094 0.07101263 0.062033135 0.031014476  0.024160631
#> [22,]  0.207800000 0.06600000 0.04910000 0.034500000 0.017200000  0.107200000
#> [23,] -0.031669763 0.01845742 0.03083832 0.025396822 0.015785316  0.056577188
#> [24,]  0.006007295 0.02248696 0.01750769 0.030316128 0.009633313  0.066280048
#> [25,]  0.199533757 0.17462885 0.09938806 0.006171233 0.002492830 -0.002590004
#>               V50         V51         V52          V53         V54         V55
#>  [1,] 0.038041389 0.032289791 0.019745381 0.0105059025 0.013300003 0.020536088
#>  [2,] 0.023590942 0.015785428 0.013687955 0.0041486301 0.007537466 0.004781475
#>  [3,] 0.018122737 0.018495277 0.015093610 0.0161142776 0.016560514 0.003704989
#>  [4,] 0.019460703 0.020320169 0.015973838 0.0135746636 0.018949852 0.013552542
#>  [5,] 0.022395489 0.013707967 0.012982607 0.0036823018 0.013360339 0.002362164
#>  [6,] 0.017532853 0.017662991 0.013705011 0.0074397505 0.004520625 0.003200389
#>  [7,] 0.015249339 0.011478751 0.007774728 0.0061907459 0.006060665 0.002595905
#>  [8,] 0.014400000 0.032900000 0.014100000 0.0019000000 0.006700000 0.009900000
#>  [9,] 0.012027480 0.023597906 0.019900878 0.0096040087 0.012244344 0.007774408
#> [10,] 0.014887454 0.022901472 0.014344444 0.0086551692 0.008712222 0.009085296
#> [11,] 0.028127531 0.009111526 0.011007023 0.0066432303 0.017994980 0.012261750
#> [12,] 0.027429813 0.022051294 0.020273785 0.0151833154 0.021064215 0.021701065
#> [13,] 0.037513920 0.032757562 0.031155894 0.0191983711 0.016504109 0.012941149
#> [14,] 0.013986839 0.024542546 0.019993625 0.0252091132 0.012018411 0.018036743
#> [15,] 0.007329695 0.008850939 0.005324455 0.0101236934 0.008097138 0.005677618
#> [16,] 0.051936461 0.045756568 0.036160225 0.0157136475 0.015163557 0.010852926
#> [17,] 0.013100946 0.010168815 0.010663692 0.0076614018 0.008558578 0.010657436
#> [18,] 0.017545919 0.004187092 0.008371179 0.0007859047 0.001795883 0.006894696
#> [19,] 0.017444812 0.023032115 0.014025430 0.0109356065 0.011729593 0.007399571
#> [20,] 0.027391542 0.017669285 0.007283407 0.0125078083 0.009922201 0.009886210
#> [21,] 0.015182884 0.010993216 0.008028486 0.0069520538 0.008679614 0.005579900
#> [22,] 0.028700000 0.002700000 0.020800000 0.0048000000 0.019900000 0.012600000
#> [23,] 0.009388515 0.003858120 0.003953835 0.0079645749 0.009115401 0.009167534
#> [24,] 0.009347629 0.004439812 0.006678939 0.0052049400 0.002999957 0.003327184
#> [25,] 0.040342856 0.003517930 0.007381239 0.0027644041 0.016245747 0.011722603
#>               V56         V57          V58         V59         V6         V60
#>  [1,] 0.012645465 0.011661805  0.013112334 0.008529729 0.11140619 0.005352040
#>  [2,] 0.008905186 0.010203039  0.010953743 0.009878692 0.12013334 0.008080667
#>  [3,] 0.012343724 0.010283886  0.011375831 0.017377379 0.10722231 0.011938953
#>  [4,] 0.011763233 0.007644177  0.013284569 0.008121179 0.16653646 0.006419111
#>  [5,] 0.010064428 0.002971711  0.009832144 0.001971448 0.04517287 0.006173283
#>  [6,] 0.005254863 0.003712403  0.001933351 0.001815829 0.11062308 0.005947221
#>  [7,] 0.005339423 0.002490051  0.003617072 0.001087526 0.09099209 0.002476339
#>  [8,] 0.004200000 0.005700000  0.005100000 0.003300000 0.10270000 0.005800000
#>  [9,] 0.007717453 0.006339531  0.008026197 0.007289057 0.11788023 0.004297774
#> [10,] 0.006326895 0.004728075  0.006611676 0.006605012 0.16012377 0.005584855
#> [11,] 0.010295141 0.007789345  0.015929135 0.014957578 0.04466553 0.007314788
#> [12,] 0.014439915 0.018125474  0.018429616 0.019244872 0.12582160 0.016254347
#> [13,] 0.007786289 0.011399790  0.009173467 0.011595125 0.11403570 0.006625754
#> [14,] 0.016865172 0.015180126  0.012401403 0.015000644 0.06727007 0.013641950
#> [15,] 0.008205197 0.005894443  0.005433019 0.006969990 0.08014192 0.007479960
#> [16,] 0.002033378 0.008454455 -0.001419144 0.004350815 0.11192858 0.003549783
#> [17,] 0.008378624 0.007907831  0.007976246 0.005132995 0.06005209 0.001994682
#> [18,] 0.003735311 0.012366432  0.013484088 0.007090579 0.07367021 0.008956454
#> [19,] 0.007225501 0.009342444  0.011093551 0.012677315 0.10289525 0.010405099
#> [20,] 0.007710423 0.006207435  0.008449761 0.010183981 0.08252738 0.007441723
#> [21,] 0.004839031 0.004752078  0.003492971 0.008717512 0.05692790 0.004406655
#> [22,] 0.002200000 0.003700000  0.003400000 0.011400000 0.25870000 0.007700000
#> [23,] 0.002722852 0.004179719  0.003285919 0.002992560 0.09486792 0.002661190
#> [24,] 0.004321864 0.005871040  0.004883599 0.002185030 0.07919106 0.002797847
#> [25,] 0.001033521 0.011169782  0.004857170 0.007833422 0.03295958 0.008904337
#>               V7         V8          V9
#>  [1,] 0.12299107 0.14022187 0.271204448
#>  [2,] 0.13521584 0.15132498 0.179643061
#>  [3,] 0.08943197 0.12227087 0.071987711
#>  [4,] 0.17674216 0.17678673 0.238339473
#>  [5,] 0.06921115 0.09858872 0.223629957
#>  [6,] 0.13451364 0.15929272 0.222733648
#>  [7,] 0.11844325 0.11199305 0.222875132
#>  [8,] 0.12870000 0.18500000 0.264700000
#>  [9,] 0.11618563 0.09923362 0.141069135
#> [10,] 0.12099214 0.18015655 0.299765833
#> [11,] 0.15330218 0.29964610 0.469153314
#> [12,] 0.20742369 0.34526312 0.419076310
#> [13,] 0.09102776 0.11868266 0.242545712
#> [14,] 0.07222342 0.03900063 0.173006529
#> [15,] 0.09266277 0.09045301 0.116018298
#> [16,] 0.16165150 0.12440749 0.079007260
#> [17,] 0.08341722 0.10596632 0.122307380
#> [18,] 0.10148835 0.11136673 0.071165567
#> [19,] 0.13721771 0.16865359 0.193090800
#> [20,] 0.08156595 0.07300505 0.009149011
#> [21,] 0.06939445 0.04747929 0.082348177
#> [22,] 0.23040000 0.20670000 0.341600000
#> [23,] 0.10453443 0.10183111 0.167553572
#> [24,] 0.05544874 0.09545904 0.124372652
#> [25,] 0.12736219 0.13023633 0.028386742
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623