Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1         V10        V11          V12         V13        V14
#>  [1,]  0.042514733  0.23763231 0.27852546  0.320273772 0.333409547 0.30109224
#>  [2,]  0.105166935  0.59080298 0.48388366  0.203167682 0.168782715 0.27812953
#>  [3,]  0.036926459  0.32085622 0.39593393  0.438306901 0.433948373 0.40575869
#>  [4,]  0.022043430  0.26434171 0.29701982  0.310832937 0.306195948 0.27371654
#>  [5,]  0.062442576  0.33758155 0.22132630  0.151542684 0.401347375 0.55251580
#>  [6,]  0.026031872  0.08254532 0.10782730  0.072105764 0.007416704 0.13491933
#>  [7,]  0.035835185  0.32808098 0.36422102  0.375180905 0.292960807 0.15643892
#>  [8,]  0.043000000  0.71060000 0.73420000  0.503300000 0.300000000 0.19510000
#>  [9,]  0.045312874  0.12254351 0.21644216  0.348965592 0.408265204 0.48227003
#> [10,]  0.007511470  0.24440843 0.32252269  0.425789814 0.554229252 0.62270547
#> [11,]  0.040384026  0.37345399 0.45990739  0.512417342 0.647567130 0.67763369
#> [12,]  0.012656853  0.14126077 0.06208622 -0.008389101 0.091791933 0.23003369
#> [13,]  0.009400000  0.50790000 0.33500000  0.083400000 0.300400000 0.39570000
#> [14,]  0.070594072 -0.03709900 0.12815219  0.176681676 0.219248442 0.31815150
#> [15,]  0.003669443  0.01975695 0.03958682  0.073759944 0.129929805 0.12551344
#> [16,]  0.014037920  0.10024415 0.08321241  0.061401605 0.092024391 0.09754043
#> [17,]  0.022548171  0.14053346 0.21785169  0.345233176 0.468876073 0.54304238
#> [18,]  0.039080165  0.04688141 0.07295881  0.054216969 0.221169825 0.29858577
#> [19,] -0.002372203  0.14318552 0.15902542  0.076073710 0.138513471 0.06009288
#> [20,]  0.017321590  0.44513973 0.34302736  0.182942561 0.349896753 0.56196171
#> [21,]  0.003776610  0.01912877 0.08565313  0.257386681 0.293137596 0.29583318
#> [22,]  0.021474100  0.08732674 0.07528436  0.160325005 0.128354922 0.09726904
#> [23,]  0.008094683  0.11796388 0.11824113  0.005241163 0.034491526 0.07261344
#> [24,]  0.029100000  0.11460000 0.09420000  0.079400000 0.025200000 0.11910000
#> [25,]  0.015351948  0.13044672 0.12135958  0.118240363 0.113706241 0.12956174
#>              V15       V16        V17        V18        V19            V2
#>  [1,] 0.27931564 0.3035386 0.30551098 0.22268239 0.26205551  0.0674442804
#>  [2,] 0.54172630 0.5344027 0.22247001 0.17126758 0.42856395  0.1060513316
#>  [3,] 0.37486531 0.4000895 0.47439576 0.56549756 0.62923037  0.0352879775
#>  [4,] 0.15912705 0.1132071 0.21937037 0.27772962 0.38577571  0.0337655426
#>  [5,] 0.23136077 0.4149404 0.58078689 0.57595850 0.80835440  0.0620107162
#>  [6,] 0.17535507 0.3627549 0.41763879 0.63913630 0.76495228 -0.0006806624
#>  [7,] 0.13066796 0.1887780 0.18818185 0.26745636 0.36086110  0.0362938592
#>  [8,] 0.27670000 0.3737000 0.25070000 0.25070000 0.32920000  0.0902000000
#>  [9,] 0.54715887 0.6150118 0.71467139 0.84955770 0.95014069  0.0721536964
#> [10,] 0.58324943 0.4772156 0.63198121 0.77412337 0.84746274  0.0411528162
#> [11,] 0.73748810 0.7479967 0.89572944 0.94246743 0.88852281  0.0367506041
#> [12,] 0.39457826 0.6076545 0.72503701 0.77718196 0.87381120  0.0027598980
#> [13,] 0.37690000 0.3828000 0.12470000 0.13630000 0.26780000  0.0611000000
#> [14,] 0.46086693 0.5308055 0.73025611 0.88170920 0.95430831  0.0523531759
#> [15,] 0.07659811 0.1065106 0.12535303 0.14006205 0.11886072  0.0026482998
#> [16,] 0.02270801 0.1133216 0.15269599 0.36154700 0.31904830  0.0067306165
#> [17,] 0.65550469 0.7113369 0.53060265 0.40616969 0.31311486  0.0384858975
#> [18,] 0.22788974 0.1119534 0.15920437 0.05615945 0.08766546  0.0768396076
#> [19,] 0.21633156 0.5030685 0.69256780 0.72091451 0.52445776 -0.0011648787
#> [20,] 0.68058390 0.9084418 0.93882859 0.87715214 0.63241091  0.0668086606
#> [21,] 0.10999714 0.1906077 0.30917797 0.37155885 0.29493560 -0.0086516694
#> [22,] 0.14012578 0.1589237 0.13773890 0.13706162 0.19031731  0.0282550129
#> [23,] 0.15239444 0.0738746 0.09149194 0.18328325 0.29588612  0.0181194435
#> [24,] 0.10450000 0.2050000 0.15560000 0.26900000 0.37840000  0.0400000000
#> [25,] 0.12849600 0.2434030 0.38307304 0.39270378 0.35143695  0.0132106661
#>             V20       V21       V22       V23        V24       V25       V26
#>  [1,] 0.3876138 0.4433766 0.4870616 0.5602634 0.59003409 0.6076853 0.6776665
#>  [2,] 0.8730098 0.8890968 0.4082713 0.1293937 0.15115232 0.2042136 0.3401853
#>  [3,] 0.7186054 0.8239710 0.8959875 0.9152259 0.96330941 0.9757056 0.9808111
#>  [4,] 0.4387015 0.4804987 0.4609804 0.4128017 0.49175698 0.6102529 0.7379554
#>  [5,] 0.8743192 0.5412177 0.6491096 0.8401640 0.67837020 0.7741551 0.8584597
#>  [6,] 0.8187866 0.8619110 0.9768756 1.0258457 0.96342309 0.9501978 0.8432648
#>  [7,] 0.3994503 0.5376161 0.7187964 0.8436167 0.90141418 0.9177551 0.9595514
#>  [8,] 0.4871000 0.6527000 0.8454000 0.9739000 1.00000000 0.6665000 0.5323000
#>  [9,] 0.9459857 0.9016742 0.8907345 0.8240854 0.72263177 0.4840936 0.3597759
#> [10,] 0.9154419 1.0157390 1.0775524 0.9275004 0.86996861 0.9182195 0.8227964
#> [11,] 0.7836731 0.6481468 0.4455621 0.1983620 0.09111623 0.2096086 0.3838333
#> [12,] 0.9337598 0.9126834 0.8942170 0.8062429 0.49716893 0.2429040 0.2358351
#> [13,] 0.9188000 0.9779000 0.3236000 0.1944000 0.18740000 0.0885000 0.3443000
#> [14,] 0.9634736 0.9224921 0.9134847 0.9445871 0.87444703 0.7214129 0.5662559
#> [15,] 0.1121138 0.1860454 0.3728289 0.4376587 0.56694702 0.6581421 0.7445259
#> [16,] 0.2968446 0.2104192 0.2108701 0.2004189 0.21294926 0.0591873 0.1939324
#> [17,] 0.4191819 0.5199242 0.5885014 0.5942985 0.78398426 0.8399195 0.7442019
#> [18,] 0.1424831 0.1406380 0.3494727 0.1035602 0.17863000 0.5716793 0.7904068
#> [19,] 0.2712228 0.3407312 0.2142385 0.3229156 0.73912995 0.9147893 0.8679348
#> [20,] 0.5002191 0.5446686 0.5562594 0.6891805 0.80282160 0.7847062 0.6959092
#> [21,] 0.5026317 0.4962022 0.4052850 0.4867266 0.67199840 0.8893625 0.8081225
#> [22,] 0.2365445 0.2425374 0.1937835 0.2553509 0.22749037 0.1441438 0.2573906
#> [23,] 0.3237265 0.4182629 0.8172700 0.9192008 0.84697013 0.6772143 0.7053298
#> [24,] 0.4024000 0.3470000 0.1395000 0.1208000 0.28270000 0.1500000 0.2626000
#> [25,] 0.4205745 0.5414924 0.6767541 0.7893423 0.81414832 0.8328887 0.9079650
#>             V27       V28       V29           V3       V30       V31        V32
#>  [1,] 0.7949244 0.8699675 0.8129705  0.068983236 0.7969972 0.7943523 0.74980629
#>  [2,] 0.2276467 0.5195445 0.5654286  0.050110253 0.7260558 0.3925709 0.41350397
#>  [3,] 0.9724472 0.9281017 0.7912346  0.030551251 0.6118312 0.4620141 0.35144107
#>  [4,] 0.9354545 1.0271881 0.9867578  0.024246915 0.7522335 0.6326225 0.51565144
#>  [5,] 0.7988435 0.6751127 0.3534573  0.070383680 0.6449344 0.7188047 0.82165662
#>  [6,] 0.6954716 0.6171636 0.5268497  0.028218326 0.4484612 0.3519187 0.23738610
#>  [7,] 0.9358179 0.9332709 0.8678682  0.047731292 0.6884171 0.6023548 0.47024076
#>  [8,] 0.4024000 0.3444000 0.4239000  0.083300000 0.4182000 0.4393000 0.11620000
#>  [9,] 0.4019108 0.3362328 0.2455861  0.071749021 0.2513997 0.1858296 0.30709694
#> [10,] 0.7033028 0.6574553 0.6579843  0.040339252 0.4385667 0.1729814 0.05097554
#> [11,] 0.4126256 0.3477028 0.1983964  0.064486253 0.1356660 0.2380716 0.28427579
#> [12,] 0.2376190 0.3930066 0.5307958 -0.001032550 0.5483974 0.4053389 0.42109476
#> [13,] 0.2953000 0.5908000 0.4564000  0.113600000 0.7334000 0.1969000 0.27900000
#> [14,] 0.4533524 0.4205944 0.4228320  0.061298761 0.2902071 0.2312288 0.30767766
#> [15,] 0.7607689 0.8546670 0.9524397  0.013020998 0.8855214 0.7150295 0.67876701
#> [16,] 0.4083237 0.6400610 0.8067892 -0.008283648 0.7627494 0.5581699 0.22082227
#> [17,] 0.6207964 0.7008505 0.8418150  0.028280971 0.7026050 0.5580463 0.48975136
#> [18,] 0.6785839 0.5357510 0.4883165  0.033709259 0.5628963 0.7349656 0.90279115
#> [19,] 0.7931279 0.8081266 0.5746687  0.006265956 0.3718362 0.4010768 0.39649159
#> [20,] 0.7589172 0.6341298 0.3014082  0.080915989 0.3565902 0.6291028 0.57887278
#> [21,] 0.6657950 0.5205590 0.4689942  0.037419551 0.3212492 0.3635190 0.50056199
#> [22,] 0.4096523 0.5561610 0.6316549  0.032611723 0.7110765 0.7152396 0.56170299
#> [23,] 0.5991244 0.6952368 0.8123750  0.039389218 0.7579648 0.5524893 0.56334007
#> [24,] 0.4468000 0.7520000 0.9036000  0.077100000 0.7812000 0.4766000 0.24830000
#> [25,] 0.9344994 0.8854413 0.8641423  0.014436322 0.8135599 0.7746262 0.71303913
#>              V33         V34         V35        V36        V37        V38
#>  [1,] 0.57230025  0.50337892  0.47566907 0.41189125 0.34858160 0.36290261
#>  [2,] 0.65469319  0.77939022  0.91558531 0.86230505 0.85290591 0.90285904
#>  [3,] 0.21693851  0.07802796  0.04161300 0.19262070 0.24512083 0.27706964
#>  [4,] 0.43690247  0.40032602  0.25231294 0.13050707 0.19888023 0.16465968
#>  [5,] 0.68566871  0.55090946  0.31919807 0.09076311 0.09898656 0.39044172
#>  [6,] 0.13489188  0.24757125  0.42487477 0.45242497 0.32833357 0.26525407
#>  [7,] 0.38035210  0.25566055  0.15593536 0.07865566 0.13087298 0.13819847
#>  [8,] 0.43360000  0.65530000  0.61720000 0.43730000 0.41180000 0.36410000
#>  [9,] 0.43293091  0.38088717  0.31101167 0.28503627 0.43366476 0.47939119
#> [10,] 0.02238363 -0.08618008 -0.01616394 0.26384398 0.25306105 0.39399176
#> [11,] 0.33299490  0.29944087  0.37664872 0.05744207 0.14435457 0.21097980
#> [12,] 0.47139309  0.36432770  0.28762842 0.20446689 0.11327216 0.17967794
#> [13,] 0.62120000  0.86810000  0.86210000 0.93800000 0.83270000 0.94800000
#> [14,] 0.35175896  0.43732382  0.39718243 0.43370823 0.55477316 0.52503879
#> [15,] 0.72751780  0.58708700  0.49077296 0.41799238 0.32588584 0.35476400
#> [16,] 0.48086825  0.45881594  0.22755054 0.31868435 0.45070925 0.56602680
#> [17,] 0.29594283  0.29117309  0.32546433 0.29505770 0.13990929 0.12135492
#> [18,] 0.86189034  0.68048650  0.86105239 1.04512811 0.98703821 0.82222033
#> [19,] 0.34284987  0.18248539  0.28281572 0.32120090 0.11107396 0.04321653
#> [20,] 0.25859379  0.36772918  0.73696444 0.66242292 0.52529771 0.56809905
#> [21,] 0.48782102  0.52031635  0.53129740 0.51506050 0.52757302 0.30737374
#> [22,] 0.66381824  0.88398962  0.95559618 0.92534697 0.71696962 0.57158167
#> [23,] 0.75452360  0.90201086  0.88488762 0.69206117 0.41963032 0.25020588
#> [24,] 0.53720000  0.62790000  0.36470000 0.45720000 0.63590000 0.64740000
#> [25,] 0.65621075  0.65510166  0.66309996 0.62424089 0.56554887 0.45930953
#>             V39            V4        V40        V41        V42        V43
#>  [1,] 0.3575416  0.0677958964 0.31489075 0.50407572 0.57327276 0.47507625
#>  [2,] 0.7759400  0.0913669353 0.52996719 0.28698319 0.25292294 0.24010745
#>  [3,] 0.3022735  0.0682444830 0.31337153 0.25615393 0.32388963 0.35109161
#>  [4,] 0.1697242  0.0602230639 0.29689842 0.32775518 0.33397166 0.21427965
#>  [5,] 0.6022616  0.1087185888 0.37758292 0.41302986 0.48294405 0.44766098
#>  [6,] 0.3617420  0.0903790134 0.39379939 0.46672895 0.55389166 0.40514468
#>  [7,] 0.1366192  0.0579697871 0.10761907 0.12413473 0.12469073 0.18970323
#>  [8,] 0.4572000  0.0813000000 0.43670000 0.29640000 0.43120000 0.41550000
#>  [9,] 0.3853829  0.0312584731 0.41220361 0.36248131 0.39620722 0.28370507
#> [10,] 0.5271235  0.0357856548 0.35427906 0.24108157 0.21586689 0.10415237
#> [11,] 0.2053323  0.1291088497 0.13278547 0.04079709 0.13557300 0.12794614
#> [12,] 0.1955784  0.0504809732 0.09352358 0.15804207 0.22076135 0.16785637
#> [13,] 0.6721000  0.1203000000 0.44360000 0.51630000 0.38090000 0.15570000
#> [14,] 0.5272714  0.0393529674 0.39807486 0.38016948 0.42173332 0.31406363
#> [15,] 0.3891951  0.0280917244 0.32348030 0.16140867 0.03792027 0.07869397
#> [16,] 0.6617497 -0.0006950151 0.53960921 0.24376304 0.22521033 0.17928444
#> [17,] 0.1524825  0.0255734912 0.14786083 0.18980634 0.12669666 0.16885870
#> [18,] 0.5774509  0.0210996122 0.31485289 0.21775604 0.19297506 0.25694640
#> [19,] 0.1003698  0.0057421329 0.12299612 0.07632177 0.06820472 0.08681574
#> [20,] 0.4484360  0.0702743247 0.31164384 0.15303099 0.16860800 0.26162200
#> [21,] 0.2436198  0.0179444515 0.45684427 0.42057626 0.42352347 0.26929028
#> [22,] 0.4915761  0.0263381483 0.56859422 0.53436619 0.48099656 0.31598694
#> [23,] 0.2523729  0.0369612522 0.32737212 0.29295619 0.22099007 0.32764461
#> [24,] 0.5520000  0.0809000000 0.32530000 0.22920000 0.06530000 0.00000000
#> [25,] 0.3029071  0.0238565290 0.22365316 0.18800904 0.10085268 0.08265008
#>              V44        V45        V46         V47          V48         V49
#>  [1,] 0.42087733 0.52397911 0.44994011  0.29271646  0.187843380  0.10950958
#>  [2,] 0.31855567 0.47838256 0.36379435  0.15364410  0.104693174  0.05706176
#>  [3,] 0.27840737 0.18252892 0.12022556  0.12454056  0.102617800  0.06005308
#>  [4,] 0.11570862 0.17128261 0.15589707  0.18795509  0.135136837  0.06432829
#>  [5,] 0.40058839 0.52546353 0.54537101  0.31121380  0.241943428  0.12898731
#>  [6,] 0.21855173 0.09603441 0.23988905  0.26502293  0.149645060  0.08045653
#>  [7,] 0.17811527 0.13747141 0.15821972  0.15567987  0.120895207  0.07472944
#>  [8,] 0.18240000 0.14870000 0.01380000  0.11640000  0.205200000  0.10690000
#>  [9,] 0.14477694 0.04070675 0.03095884  0.01380101  0.013678160  0.01993712
#> [10,] 0.17514511 0.15428517 0.10328983  0.12275136  0.117793071  0.08771131
#> [11,] 0.04220253 0.09974644 0.08482474  0.05164980  0.068350213  0.05025818
#> [12,] 0.21296194 0.18985213 0.07627948  0.05638760  0.058733934  0.04055261
#> [13,] 0.14490000 0.26620000 0.18060000  0.16990000  0.255900000  0.11290000
#> [14,] 0.31824461 0.18607925 0.08259893  0.04967431  0.003418569  0.02893416
#> [15,] 0.10591954 0.11193937 0.04397769  0.02161675  0.038541281  0.01687236
#> [16,] 0.08602140 0.04543637 0.03068250  0.09565919  0.093647905  0.03045123
#> [17,] 0.15862925 0.10705645 0.15697495  0.11387214  0.033573758  0.02883837
#> [18,] 0.29491983 0.26570681 0.19892466  0.11515358  0.094987288  0.04593549
#> [19,] 0.08731048 0.07873542 0.08266969  0.02672399  0.037677214  0.01086836
#> [20,] 0.26431923 0.21087502 0.07324325  0.07364841  0.057818531  0.02746724
#> [21,] 0.18471784 0.04418268 0.02075900 -0.04023301 -0.019532377 -0.01244324
#> [22,] 0.21615134 0.16561683 0.13099921  0.12535304  0.090002212  0.04790835
#> [23,] 0.23664094 0.11227187 0.05016108  0.04301865  0.039581641  0.01296323
#> [24,] 0.00000000 0.00000000 0.00000000  0.00000000  0.000000000  0.00000000
#> [25,] 0.05199858 0.04237326 0.05449788  0.03253724  0.027917736  0.01089215
#>               V5          V50           V51          V52           V53
#>  [1,] 0.07530815  0.040109892  0.0350211263  0.027784744  0.0145250161
#>  [2,] 0.07948659  0.031167253  0.0183892832  0.021740574  0.0158731374
#>  [3,] 0.11258382  0.023832957  0.0207284979  0.009100860  0.0143595143
#>  [4,] 0.12591858  0.008191442  0.0207452741  0.013673887  0.0102296802
#>  [5,] 0.23737405  0.080206948  0.0231051717  0.020195463  0.0340686883
#>  [6,] 0.10741406  0.043739948  0.0300633385  0.025241306  0.0309151234
#>  [7,] 0.07750621  0.021339915  0.0198703794  0.015316523  0.0072602566
#>  [8,] 0.01650000  0.019900000  0.0208000000  0.017600000  0.0197000000
#>  [9,] 0.06788062  0.018013377  0.0102470921  0.020215316  0.0174892788
#> [10,] 0.17091443  0.027508138  0.0286923821  0.018568229  0.0128059452
#> [11,] 0.09481423  0.030950933  0.0311972906  0.029992600  0.0226063418
#> [12,] 0.08002586  0.003756847  0.0152078127  0.012630086  0.0024263806
#> [13,] 0.04030000  0.020100000  0.0480000000  0.023400000  0.0175000000
#> [14,] 0.11924253  0.006976397  0.0001822096 -0.002288212 -0.0001362383
#> [15,] 0.02375476  0.005442094  0.0071670683  0.007040038  0.0126980766
#> [16,] 0.02922120  0.014016991  0.0073129272  0.021801211  0.0256614785
#> [17,] 0.07501968  0.022794160  0.0138786448  0.010158216  0.0097330419
#> [18,] 0.02643267 -0.001752631  0.0091801030  0.016094710  0.0268394458
#> [19,] 0.05787497  0.017874321  0.0045613333  0.023107571  0.0088251135
#> [20,] 0.10671974  0.024727168  0.0052621910  0.019854042  0.0065566007
#> [21,] 0.02457125 -0.002026979 -0.0031212738 -0.002860522  0.0088155750
#> [22,] 0.03976074  0.021994077  0.0170830810  0.009002918  0.0123858723
#> [23,] 0.06906237  0.015896257  0.0159063201  0.014260418  0.0195077100
#> [24,] 0.05210000  0.000000000  0.0000000000  0.005600000  0.0237000000
#> [25,] 0.04614058  0.012645829  0.0063886280  0.007251867  0.0056205792
#>               V54         V55         V56          V57          V58         V59
#>  [1,] 0.014964874 0.013688256 0.008887365 0.0093632109 0.0095677335 0.009943545
#>  [2,] 0.018928188 0.021419677 0.010294367 0.0119960808 0.0129552761 0.008246583
#>  [3,] 0.014096191 0.005736554 0.010795314 0.0068399262 0.0151152598 0.008014878
#>  [4,] 0.006189305 0.001418510 0.005323995 0.0002616107 0.0041085203 0.004070465
#>  [5,] 0.022609527 0.042460225 0.037141566 0.0337749224 0.0419286473 0.023310397
#>  [6,] 0.021728282 0.002264339 0.006773032 0.0140953320 0.0099802969 0.009087143
#>  [7,] 0.008091647 0.005413816 0.006415136 0.0070422580 0.0077073426 0.006779338
#>  [8,] 0.021000000 0.014100000 0.004900000 0.0027000000 0.0162000000 0.005900000
#>  [9,] 0.022726665 0.011489891 0.015781980 0.0023326943 0.0074641436 0.006855859
#> [10,] 0.035685259 0.023507655 0.003209101 0.0081392965 0.0077743216 0.011578618
#> [11,] 0.011320800 0.017637845 0.013620675 0.0003946449 0.0310205883 0.017008472
#> [12,] 0.003041313 0.002479902 0.002444242 0.0012711458 0.0048977508 0.003475358
#> [13,] 0.035200000 0.015800000 0.032600000 0.0201000000 0.0168000000 0.024500000
#> [14,] 0.028256215 0.021876757 0.010735890 0.0124719031 0.0125456087 0.014452917
#> [15,] 0.004957348 0.005396743 0.008109026 0.0055303749 0.0007328913 0.005334548
#> [16,] 0.014467704 0.023490186 0.004200475 0.0014842199 0.0045165156 0.008580631
#> [17,] 0.005735077 0.007042231 0.010957571 0.0077005495 0.0046553968 0.004597413
#> [18,] 0.003155497 0.023397688 0.011056002 0.0030064819 0.0103956027 0.010587424
#> [19,] 0.013145206 0.012018392 0.005760096 0.0058924532 0.0020722200 0.005550893
#> [20,] 0.018685255 0.013752587 0.005475964 0.0076834766 0.0057751705 0.011113393
#> [21,] 0.013258061 0.014170386 0.011529361 0.0110292102 0.0046175836 0.011720082
#> [22,] 0.012889759 0.009526032 0.007000453 0.0091479780 0.0110116848 0.014424108
#> [23,] 0.015245747 0.005796599 0.013300532 0.0219670772 0.0081520558 0.007638470
#> [24,] 0.020400000 0.005000000 0.013700000 0.0164000000 0.0081000000 0.013900000
#> [25,] 0.006138413 0.006817265 0.004916774 0.0060387261 0.0061268724 0.004468719
#>               V6          V60         V7         V8          V9
#>  [1,] 0.12394168  0.007369115 0.10985777 0.14974084  0.23826070
#>  [2,] 0.10935102  0.012909661 0.25192832 0.37967416  0.54532500
#>  [3,] 0.16463448  0.006293706 0.15582658 0.19230995  0.26007813
#>  [4,] 0.12171288  0.003262177 0.08866466 0.10982887  0.22380906
#>  [5,] 0.12629491  0.009406003 0.17717268 0.10225949  0.24569110
#>  [6,] 0.06366699  0.008350378 0.14495220 0.13017946  0.03269380
#>  [7,] 0.12167994  0.006489916 0.14101793 0.17569844  0.23796210
#>  [8,] 0.02770000  0.002100000 0.05690000 0.20570000  0.38870000
#>  [9,] 0.09157128  0.008939939 0.06606730 0.07265015  0.08118807
#> [10,] 0.23964816  0.009113532 0.21580176 0.20211097  0.24271385
#> [11,] 0.11504104  0.015653986 0.16754547 0.09853086  0.21415038
#> [12,] 0.13975055 -0.001368110 0.13578811 0.12614394  0.17598049
#> [13,] 0.12270000  0.015400000 0.24950000 0.45660000  0.65870000
#> [14,] 0.07483370  0.011094504 0.05614121 0.02835846  0.08538728
#> [15,] 0.02251224  0.004134638 0.02336431 0.01574155  0.01597031
#> [16,] 0.04257048  0.007976092 0.03074609 0.07656165  0.13789455
#> [17,] 0.12338972  0.003078521 0.14002455 0.15102464  0.16123024
#> [18,] 0.07659572  0.005709243 0.10153299 0.04666631 -0.02141925
#> [19,] 0.05650025  0.002127114 0.01375595 0.01817171  0.03936771
#> [20,] 0.23863036  0.007598181 0.22953944 0.23040208  0.36330248
#> [21,] 0.06543677  0.010290980 0.08710767 0.13172161  0.05257882
#> [22,] 0.07358076  0.003419984 0.07855044 0.07423748  0.04329671
#> [23,] 0.06912501  0.012931041 0.10210801 0.07973298  0.02323641
#> [24,] 0.10510000  0.011100000 0.01450000 0.06740000  0.12940000
#> [25,] 0.09015616  0.005255608 0.10134453 0.09800072  0.11918390
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841