Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.022366912  0.21519924 0.30314552 0.29545123 0.30480615 0.29392273
#>  [2,] 0.023109746  0.21138758 0.26235982 0.29343335 0.29735748 0.27371769
#>  [3,] 0.049834809  0.15432019 0.21983855 0.32372539 0.43589467 0.53125125
#>  [4,] 0.033887437  0.39997546 0.47411328 0.47182338 0.46451425 0.36481997
#>  [5,] 0.027692864 -0.02336679 0.06586216 0.14029200 0.13032355 0.27104353
#>  [6,] 0.035571409  0.25167641 0.29101279 0.28377393 0.19897663 0.08399002
#>  [7,] 0.017972226  0.12814529 0.15963937 0.15538579 0.12561079 0.16091724
#>  [8,] 0.009612898  0.24447625 0.25837318 0.19440170 0.37496563 0.54466289
#>  [9,] 0.019878722  0.40304166 0.37701211 0.40442654 0.43102214 0.48535194
#> [10,] 0.026747614  0.26917395 0.32985317 0.37114361 0.37002862 0.42704893
#> [11,] 0.046720876  0.22303392 0.19690557 0.22234513 0.28418562 0.26043290
#> [12,] 0.018469448  0.16705495 0.28527689 0.38731428 0.42476491 0.45779746
#> [13,] 0.011816729  0.17261167 0.24314140 0.27000523 0.27830724 0.24897666
#> [14,] 0.055414208  0.47737033 0.37636098 0.24764244 0.21928173 0.26211142
#> [15,] 0.025075075  0.61851583 0.63613575 0.71553843 0.56269632 0.53407302
#> [16,] 0.032012705  0.33197980 0.26519196 0.12152002 0.21660629 0.46866161
#> [17,] 0.014707080  0.10546240 0.15786350 0.17386646 0.14244229 0.20347106
#> [18,] 0.012226001  0.04790699 0.11454881 0.05466148 0.12006085 0.19633496
#> [19,] 0.013122239  0.18060526 0.12810876 0.06203417 0.06438701 0.08305694
#> [20,] 0.011726834  0.10808855 0.10079094 0.09375842 0.10332174 0.11844048
#> [21,] 0.011050891  0.09044284 0.06375592 0.10663479 0.20940997 0.20126294
#> [22,] 0.038151638  0.07273635 0.09896097 0.22286523 0.18069834 0.14444869
#> [23,] 0.015709939  0.11720288 0.13043070 0.11271821 0.09410837 0.11428561
#> [24,] 0.019269547  0.16102897 0.23933052 0.17275122 0.14668737 0.17137137
#> [25,] 0.019009266  0.10370326 0.08270461 0.08594871 0.10943162 0.03935153
#>               V15        V16        V17       V18       V19           V2
#>  [1,] 0.258426988 0.16776384 0.23642498 0.2862360 0.3679468 0.0293242051
#>  [2,] 0.259581765 0.29291664 0.34739914 0.4675343 0.6269185 0.0430224360
#>  [3,] 0.622446669 0.62654199 0.71630868 0.7943110 0.8649385 0.0507739448
#>  [4,] 0.256989896 0.28330155 0.30246919 0.3309116 0.4820801 0.0540327829
#>  [5,] 0.357438874 0.58247137 0.69741982 0.7414161 0.8576750 0.0320101366
#>  [6,] 0.092177522 0.16869345 0.20631669 0.2492184 0.2580759 0.0340423382
#>  [7,] 0.170364686 0.36191947 0.45250741 0.6061090 0.7007554 0.0029329978
#>  [8,] 0.557247209 0.62363640 0.61173815 0.7521024 0.9108660 0.0191249616
#>  [9,] 0.466572779 0.60533384 0.80489587 0.8162579 0.8152165 0.1059662285
#> [10,] 0.463084682 0.49305706 0.56430411 0.6580903 0.7209157 0.0130732759
#> [11,] 0.253210274 0.26567022 0.26039238 0.1365373 0.2607037 0.0526965159
#> [12,] 0.582535737 0.62867535 0.80293555 0.9522210 1.0073213 0.0238861184
#> [13,] 0.190825682 0.13358148 0.18197176 0.2439188 0.2683725 0.0231731371
#> [14,] 0.290306435 0.32830418 0.21536694 0.2258560 0.2881504 0.0612357236
#> [15,] 0.652084380 0.69811103 0.68312892 0.7626655 0.9005785 0.0579336825
#> [16,] 0.585311211 0.64895602 0.71336944 0.6118610 0.3937996 0.0706889460
#> [17,] 0.272552731 0.30569068 0.36844053 0.3583188 0.3080170 0.0234721879
#> [18,] 0.184828784 0.20948090 0.19938626 0.1659451 0.1249104 0.0026020587
#> [19,] 0.127455489 0.22244986 0.27020422 0.2640204 0.3757048 0.0009589389
#> [20,] 0.168825778 0.17966020 0.14499052 0.1492990 0.1080524 0.0150144934
#> [21,] 0.195867300 0.25658864 0.28481388 0.3442016 0.4708610 0.0189116576
#> [22,] 0.155335446 0.18322614 0.21212606 0.2110931 0.2426750 0.0498325543
#> [23,] 0.101016564 0.10993433 0.19487218 0.2710823 0.2948401 0.0173566603
#> [24,] 0.178431589 0.20601817 0.24630232 0.1102742 0.3930152 0.0050564120
#> [25,] 0.002538364 0.02113114 0.05743917 0.1974900 0.3499384 0.0026552286
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.4251920 0.4999536 0.4684185 0.4800321 0.5728375 0.6899339 0.7969353
#>  [2,] 0.7077649 0.8032179 0.8255860 0.8657575 0.9489620 0.9540962 0.9779340
#>  [3,] 0.8424295 0.9334984 0.9073725 0.8600192 0.9095040 0.7517902 0.5994431
#>  [4,] 0.6823773 0.7464172 0.7913493 0.7840970 0.7659828 0.7486689 0.7959246
#>  [5,] 0.8738417 0.8056268 0.8281130 0.7975904 0.6671792 0.5502236 0.4648947
#>  [6,] 0.3271218 0.4828773 0.6643215 0.7565935 0.8037328 0.9124988 0.9811263
#>  [7,] 0.7390858 0.7980269 0.9174409 0.9976738 0.9642611 0.9407493 0.7986843
#>  [8,] 0.9876799 1.0501093 1.0626972 1.0026449 0.8118074 0.6041141 0.6075447
#>  [9,] 0.8469048 0.7886283 0.8442567 1.0365490 1.1075503 1.0069386 0.9746416
#> [10,] 0.7831142 0.8565453 0.9254852 0.9467666 0.9915782 0.9703223 0.9589932
#> [11,] 0.3495098 0.3869626 0.4001784 0.4414346 0.4366766 0.4661337 0.4748113
#> [12,] 0.9745656 0.8893702 0.7612439 0.6366477 0.5797856 0.4519300 0.3079778
#> [13,] 0.3210443 0.3785808 0.3914608 0.5110784 0.6896705 0.7752851 0.8216167
#> [14,] 0.5356398 0.4904105 0.2468039 0.2096964 0.2938061 0.4037765 0.5762931
#> [15,] 0.8610670 0.7932135 0.6801939 0.4362903 0.3720438 0.5395193 0.2379593
#> [16,] 0.3724452 0.5228747 0.7249271 0.8222053 0.8475842 0.7031584 0.6306831
#> [17,] 0.3377250 0.3256212 0.4319284 0.5448321 0.6654719 0.7170002 0.7011971
#> [18,] 0.1066036 0.1426667 0.3194665 0.4641917 0.5884294 0.6762212 0.8080817
#> [19,] 0.5698297 0.6675536 0.8846633 1.0325692 0.8932502 0.8224276 0.9020597
#> [20,] 0.1305609 0.2167215 0.2630332 0.3941457 0.6068037 0.8368324 0.9756211
#> [21,] 0.5607489 0.5885186 0.6139458 0.5478873 0.5713250 0.7558474 0.8550363
#> [22,] 0.1837321 0.2014989 0.1136055 0.2337678 0.2632331 0.1943045 0.2905035
#> [23,] 0.4581862 0.5883318 0.7530624 0.8824515 0.8690315 0.8475997 0.8814900
#> [24,] 0.5926018 0.3822681 0.4028016 1.0323141 0.9099741 0.4553482 0.5247561
#> [25,] 0.3668989 0.3292003 0.4121259 0.5764184 0.7156991 0.8151941 0.8296162
#>             V27       V28       V29           V3       V30       V31        V32
#>  [1,] 0.9273579 0.9685711 0.8986950  0.030017523 0.7550926 0.6461359 0.52774161
#>  [2,] 0.9702145 0.9719155 0.8761866  0.031072949 0.6667472 0.4694567 0.32555482
#>  [3,] 0.5088424 0.3365156 0.3363362  0.062345401 0.3624821 0.2326473 0.07562629
#>  [4,] 0.8386420 0.8002226 0.6390296  0.054471706 0.5264458 0.4911526 0.49082763
#>  [5,] 0.5002133 0.5176531 0.6164897  0.026255985 0.5176229 0.2821412 0.25165613
#>  [6,] 0.9899264 0.9220417 0.7970289  0.034686027 0.7165003 0.6814201 0.52648933
#>  [7,] 0.6282317 0.5271639 0.4077647  0.022722591 0.4486577 0.4178505 0.33528708
#>  [8,] 0.6700338 0.6914240 0.5802360  0.034172377 0.4001401 0.2620361 0.17680029
#>  [9,] 0.9861228 0.7446521 0.5966388  0.120751194 0.5121293 0.2752806 0.20617726
#> [10,] 0.9571634 0.8774305 0.7631295  0.028994346 0.6151162 0.4371313 0.34138912
#> [11,] 0.5346256 0.7409076 0.8459513  0.068234210 0.8677203 0.8362766 0.69695322
#> [12,] 0.2738922 0.3056211 0.2861420  0.048636068 0.2915321 0.2828081 0.34129561
#> [13,] 0.9284718 1.0116052 0.8911046  0.024458091 0.6564955 0.6255022 0.46392733
#> [14,] 0.5870475 0.7472119 0.7348677  0.045668606 0.8673350 0.6750309 0.65542267
#> [15,] 0.5129889 0.8603010 0.6017635  0.111491513 0.8518218 0.8601184 0.50727822
#> [16,] 0.6775970 0.5910632 0.4150794  0.096301227 0.4235769 0.4636556 0.30909514
#> [17,] 0.6637435 0.7576212 0.8287292  0.026041402 0.7435252 0.6484548 0.56709713
#> [18,] 0.8545605 0.9180962 0.8390626  0.019254674 0.7306128 0.6345180 0.52099334
#> [19,] 0.9207604 0.9058442 0.8234173 -0.002513991 0.7159984 0.6246747 0.53220182
#> [20,] 0.8323082 0.6040049 0.4929331  0.011691488 0.3767653 0.2895081 0.14650189
#> [21,] 0.9534744 0.8552534 0.6168637  0.012371751 0.4306417 0.4384368 0.40523333
#> [22,] 0.4442093 0.6168379 0.6771279  0.036682565 0.6990424 0.7632514 0.66140648
#> [23,] 0.8611162 0.8316260 0.8351224  0.021359365 0.7506462 0.6808047 0.62412936
#> [24,] 0.6235870 0.7373470 0.5827326  0.054403055 0.3937123 0.2658316 0.18880577
#> [25,] 0.7443760 0.8212433 0.9846927  0.010838177 0.9827395 0.9359684 0.88988524
#>              V33        V34        V35        V36        V37        V38
#>  [1,] 0.45834473 0.41019403 0.28661528 0.20819624 0.23285791 0.18046333
#>  [2,] 0.25599440 0.20290627 0.20056661 0.13681765 0.09285155 0.13765165
#>  [3,] 0.31365705 0.42276702 0.44832336 0.42459618 0.33748241 0.38543721
#>  [4,] 0.47403426 0.43992138 0.28752904 0.34694445 0.31380863 0.35750717
#>  [5,] 0.29856855 0.32493328 0.41147725 0.38391019 0.35120386 0.43442953
#>  [6,] 0.36571406 0.16409884 0.09802684 0.05268924 0.11946538 0.15976530
#>  [7,] 0.31695109 0.38935851 0.43957886 0.37305135 0.22597997 0.23107571
#>  [8,] 0.03866419 0.05076933 0.22990878 0.33640272 0.23608992 0.22162400
#>  [9,] 0.09387162 0.29316337 0.41274794 0.28151413 0.24672557 0.28943640
#> [10,] 0.18814113 0.03177649 0.05182348 0.15434892 0.22699256 0.25031855
#> [11,] 0.52570629 0.43920018 0.38819312 0.25534283 0.21913805 0.34688416
#> [12,] 0.37888550 0.37532831 0.39112934 0.17993143 0.21117384 0.39137427
#> [13,] 0.31230773 0.20955455 0.16010850 0.08107828 0.18774227 0.15417807
#> [14,] 0.67662742 0.69996389 0.76165549 0.77942989 0.78735746 0.77042307
#> [15,] 0.17790368 0.25953570 0.41525096 0.29549874 0.60959729 0.67263429
#> [16,] 0.24050564 0.12812801 0.35977676 0.61169062 0.65784845 0.35724365
#> [17,] 0.65913489 0.84424969 0.82188471 0.73417986 0.57901061 0.44679709
#> [18,] 0.46704145 0.52785809 0.51921144 0.39768277 0.13680449 0.04917077
#> [19,] 0.50641235 0.55981550 0.58517551 0.51138909 0.49255616 0.33704173
#> [20,] 0.17294836 0.20604407 0.24297529 0.18511428 0.06497809 0.06563576
#> [21,] 0.28583754 0.23692560 0.25917179 0.27418804 0.32169605 0.37007463
#> [22,] 0.70464373 0.83341558 0.92138903 0.93343727 0.76672820 0.66273446
#> [23,] 0.62650554 0.64369569 0.62736081 0.54381107 0.42895715 0.30176604
#> [24,] 0.12594528 0.32879149 0.28836253 0.35149817 0.28080864 0.20322835
#> [25,] 0.93599502 0.91059238 0.74212971 0.52221972 0.32272301 0.18392910
#>              V39          V4        V40        V41          V42         V43
#>  [1,] 0.20389967 0.067338560 0.27791435 0.32763071 0.3592055925  0.25556529
#>  [2,] 0.16553788 0.023914662 0.18094370 0.11041117 0.1146671657  0.18952505
#>  [3,] 0.39104676 0.079580498 0.33693357 0.37968789 0.4011956585  0.31379587
#>  [4,] 0.50198011 0.055135467 0.43705106 0.29479735 0.4001445384  0.40400800
#>  [5,] 0.40439433 0.056506313 0.44107855 0.34355904 0.3508603131  0.49749394
#>  [6,] 0.14478275 0.041329894 0.07463295 0.09119372 0.1163858840  0.15634203
#>  [7,] 0.29804701 0.069354705 0.37380110 0.40605332 0.3952654967  0.29692531
#>  [8,] 0.27447755 0.052419287 0.21427488 0.27907352 0.3287879148  0.30027681
#>  [9,] 0.23431523 0.048484891 0.20684188 0.37051849 0.3207783397 -0.01171856
#> [10,] 0.26557616 0.083699903 0.30791370 0.29149473 0.3282801707  0.27580939
#> [11,] 0.39823068 0.061427550 0.37214083 0.53894867 0.5244045891  0.42386207
#> [12,] 0.30459803 0.105831998 0.20851281 0.16386234 0.1843607748  0.18666711
#> [13,] 0.24615482 0.049883935 0.21807058 0.21882650 0.2959278512  0.15258244
#> [14,] 0.65354337 0.077538035 0.53923772 0.42879021 0.3585148487  0.35683960
#> [15,] 0.53275742 0.107070679 0.46909834 0.46489814 0.2541363841  0.20789536
#> [16,] 0.24704728 0.087342412 0.41766323 0.44759498 0.3694449932  0.22830165
#> [17,] 0.30952187 0.042797807 0.31064325 0.28656200 0.2879834871  0.26798573
#> [18,] 0.08882048 0.020249740 0.11393372 0.13285449 0.1749402818  0.26194366
#> [19,] 0.17773921 0.028999048 0.07116029 0.08059299 0.0119930524  0.06778362
#> [20,] 0.15861279 0.013111041 0.22398297 0.28730043 0.2810156785  0.20836765
#> [21,] 0.30262713 0.006715496 0.23274143 0.19128122 0.1933654774  0.15269649
#> [22,] 0.59798394 0.024036827 0.57696301 0.53018432 0.4757500446  0.35150493
#> [23,] 0.17524347 0.040333276 0.22445497 0.21112866 0.1657277560  0.15722492
#> [24,] 0.24456870 0.050691268 0.11707100 0.02068779 0.0009901797  0.06419039
#> [25,] 0.09687296 0.008549240 0.26125820 0.31340688 0.3480343349  0.32459367
#>              V44         V45         V46        V47         V48        V49
#>  [1,] 0.17292224  0.19320786 0.188441138 0.18628124 0.151338866 0.06794921
#>  [2,] 0.19222840  0.19077625 0.173623585 0.13084574 0.082115724 0.04931445
#>  [3,] 0.28432673  0.23201385 0.045333485 0.07821564 0.035319727 0.01607153
#>  [4,] 0.36019011  0.39062402 0.222773534 0.14406663 0.146150928 0.08930792
#>  [5,] 0.34845742  0.11305202 0.120863596 0.16425986 0.089785537 0.04685592
#>  [6,] 0.14888483  0.10891560 0.095570355 0.12222624 0.101431219 0.05689396
#>  [7,] 0.16760254  0.12927348 0.197861772 0.19823361 0.123652264 0.06719212
#>  [8,] 0.24456482  0.13455794 0.098075315 0.08880583 0.115688225 0.08277617
#>  [9,] 0.12813233  0.18193882 0.164428565 0.08968660 0.062273383 0.01154976
#> [10,] 0.17448044  0.10123794 0.119832223 0.13711983 0.100713884 0.05237579
#> [11,] 0.42180702  0.53835688 0.536270539 0.39555133 0.242012471 0.12884909
#> [12,] 0.08171530  0.06715035 0.034859448 0.03501793 0.044162297 0.02134039
#> [13,] 0.07361378  0.08196793 0.186384456 0.19644316 0.129764165 0.04694257
#> [14,] 0.42611609  0.48225648 0.349776317 0.18646191 0.181104873 0.09404382
#> [15,] 0.21900419  0.20559895 0.009795833 0.13603571 0.072859620 0.01162154
#> [16,] 0.15786513  0.16710893 0.143698441 0.13903710 0.128260126 0.08054223
#> [17,] 0.20690770  0.13525955 0.114067701 0.10137541 0.068584632 0.03094673
#> [18,] 0.19366484  0.06680371 0.058520456 0.03201346 0.029732474 0.02673874
#> [19,] 0.01059977 -0.03051334 0.016339061 0.02873585 0.031761522 0.01686451
#> [20,] 0.12476264  0.09951910 0.103797820 0.05301982 0.034346323 0.02892935
#> [21,] 0.11962483  0.10036021 0.065450312 0.07562781 0.060255579 0.02433621
#> [22,] 0.25658223  0.16793476 0.105726530 0.10574187 0.091745808 0.05553240
#> [23,] 0.08827311  0.05331584 0.037995715 0.03474162 0.038641977 0.01188114
#> [24,] 0.09723917  0.05469925 0.007766335 0.01727874 0.008839423 0.01068502
#> [25,] 0.31498823  0.28465372 0.197630957 0.09372375 0.033076142 0.02591500
#>               V5         V50          V51         V52         V53           V54
#>  [1,] 0.08743652 0.014956311 0.0202932311 0.014343759 0.006223701  0.0078847992
#>  [2,] 0.08767359 0.015123232 0.0213544374 0.012400632 0.012344398  0.0058547512
#>  [3,] 0.04881356 0.006694946 0.0185278442 0.003905289 0.014217184  0.0146259582
#>  [4,] 0.07236682 0.028675761 0.0221640362 0.016464274 0.010132764  0.0116648921
#>  [5,] 0.05416893 0.026537727 0.0122670339 0.006440198 0.019330157  0.0177544323
#>  [6,] 0.03208362 0.022749449 0.0108803774 0.010991903 0.005883090  0.0107802275
#>  [7,] 0.09658046 0.030074956 0.0205504248 0.016473580 0.016217166  0.0114236887
#>  [8,] 0.15603777 0.016160737 0.0243423394 0.025771971 0.017041521  0.0265624819
#>  [9,] 0.13799422 0.015483642 0.0171959453 0.021344849 0.023671662  0.0324343087
#> [10,] 0.12512157 0.022094095 0.0150149988 0.009455061 0.010060932  0.0163267005
#> [11,] 0.05457111 0.043295443 0.0395124196 0.032744521 0.015419229  0.0159985506
#> [12,] 0.12734311 0.013007268 0.0131421434 0.022886750 0.014335309  0.0137654961
#> [13,] 0.08803542 0.007958233 0.0188886690 0.009642675 0.006048634  0.0064283585
#> [14,] 0.07563822 0.033898836 0.0231247985 0.023438291 0.014907403  0.0151154695
#> [15,] 0.09664153 0.009826614 0.0029591137 0.023059384 0.016122759  0.0090760226
#> [16,] 0.07807979 0.018470713 0.0218681185 0.012513740 0.009123600  0.0124609023
#> [17,] 0.06702760 0.018485459 0.0220727328 0.011682376 0.010116884  0.0107903389
#> [18,] 0.01819733 0.007513464 0.0145867396 0.012361869 0.013227337  0.0028540616
#> [19,] 0.05947707 0.008236885 0.0009500694 0.003030969 0.008661844  0.0093375424
#> [20,] 0.03092388 0.010588955 0.0053698429 0.008311377 0.003494007  0.0040751470
#> [21,] 0.01764131 0.014932399 0.0072637775 0.006304690 0.008390828  0.0128833508
#> [22,] 0.06011058 0.016076861 0.0089017749 0.008936720 0.016265154  0.0114354051
#> [23,] 0.06956074 0.010112073 0.0069858168 0.005568186 0.006815793  0.0045297740
#> [24,] 0.06162627 0.009182170 0.0331713560 0.044585426 0.016155625 -0.0009161614
#> [25,] 0.01622784 0.011759954 0.0074580323 0.008413846 0.006826840  0.0067188049
#>               V55          V56         V57         V58          V59          V6
#>  [1,] 0.004827295 0.0057478006 0.003800990 0.007251401 0.0060901357 0.092727673
#>  [2,] 0.006468143 0.0052510481 0.004584007 0.006333637 0.0050543117 0.124301650
#>  [3,] 0.004986240 0.0180561827 0.017840219 0.013505368 0.0254030402 0.138567634
#>  [4,] 0.014102265 0.0077068224 0.007640115 0.010149773 0.0086263959 0.116421067
#>  [5,] 0.000963046 0.0141595487 0.003433900 0.002693421 0.0143793731 0.027398794
#>  [6,] 0.005086159 0.0058592569 0.006931389 0.009381124 0.0069706693 0.124259027
#>  [7,] 0.003782756 0.0059746048 0.006650583 0.007851278 0.0084949685 0.050885115
#>  [8,] 0.015390866 0.0040142785 0.005866950 0.022921055 0.0122396673 0.242614326
#>  [9,] 0.016109533 0.0142621699 0.013726701 0.011199663 0.0002688929 0.192219336
#> [10,] 0.004157403 0.0139780564 0.006455040 0.018105514 0.0083231914 0.159143693
#> [11,] 0.013357639 0.0107284701 0.012394313 0.011280664 0.0136862437 0.088946792
#> [12,] 0.026064293 0.0196757430 0.003866012 0.015863098 0.0147025751 0.058361224
#> [13,] 0.006478670 0.0062053765 0.003874072 0.003118462 0.0022586170 0.109860125
#> [14,] 0.011930029 0.0115973104 0.011243695 0.007809929 0.0060211775 0.096741358
#> [15,] 0.018010653 0.0243893473 0.031972844 0.016593879 0.0096491378 0.232376128
#> [16,] 0.009988205 0.0056654193 0.009727796 0.008030331 0.0090295512 0.101192516
#> [17,] 0.008269596 0.0069204229 0.010358932 0.004814385 0.0088812770 0.108349108
#> [18,] 0.003987463 0.0018887982 0.011240801 0.005780267 0.0047271328 0.049116135
#> [19,] 0.009428866 0.0022097113 0.003595674 0.001545350 0.0020006978 0.102915950
#> [20,] 0.006180964 0.0034540655 0.003026210 0.003089881 0.0013789457 0.043296773
#> [21,] 0.009524866 0.0087622396 0.006700397 0.008512589 0.0046603525 0.050800160
#> [22,] 0.008552707 0.0069827873 0.009415364 0.009735742 0.0085432076 0.100046973
#> [23,] 0.004305265 0.0055654019 0.008068648 0.005931799 0.0033413388 0.100702310
#> [24,] 0.010136203 0.0093435722 0.008207674 0.019577624 0.0115310460 0.160788347
#> [25,] 0.001253405 0.0007526526 0.001094880 0.002217037 0.0047832419 0.009607278
#>               V60           V7         V8          V9
#>  [1,] 0.002935183  0.103483101 0.09060551  0.17350518
#>  [2,] 0.004378680  0.121172013 0.14578224  0.18872868
#>  [3,] 0.019013512  0.216820517 0.13623664  0.06803552
#>  [4,] 0.005680291  0.124012121 0.19463542  0.29130166
#>  [5,] 0.010442788 -0.017167887 0.05719269 -0.01080598
#>  [6,] 0.005821010  0.145768824 0.13538861  0.19416788
#>  [7,] 0.007808872  0.114214012 0.13398031  0.13100416
#>  [8,] 0.001089291  0.179658428 0.14757044  0.18755327
#>  [9,] 0.002964111  0.173004877 0.24383272  0.37359343
#> [10,] 0.006655969  0.149247816 0.14648185  0.21912951
#> [11,] 0.007085115  0.086380873 0.12782932  0.19217770
#> [12,] 0.009486299  0.089987800 0.06739247  0.08211629
#> [13,] 0.001263165  0.082920209 0.09598636  0.14258436
#> [14,] 0.006851363  0.161990780 0.26169029  0.45697108
#> [15,] 0.007750998  0.246636696 0.37942868  0.55965938
#> [16,] 0.008045221  0.105345975 0.18341801  0.25543940
#> [17,] 0.008355567  0.133933760 0.10284160  0.09001989
#> [18,] 0.006986536  0.004377912 0.03895888  0.07819722
#> [19,] 0.002702191  0.110467388 0.09454603  0.16129875
#> [20,] 0.002961767  0.061214201 0.06256539  0.10612069
#> [21,] 0.005571320  0.071414660 0.05081564  0.07332206
#> [22,] 0.006136100  0.093156363 0.06176949  0.05309262
#> [23,] 0.005158812  0.090076802 0.10961794  0.11153325
#> [24,] 0.006563018  0.141048837 0.21983335  0.15670126
#> [25,] 0.003198456  0.058236025 0.03201655  0.04468335
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3333333