Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.021718341 0.27094693 0.31736836 0.37584914 0.29978215 0.23158321
#>  [2,] 0.060231650 0.22386376 0.30810551 0.40047403 0.49631129 0.58309898
#>  [3,] 0.101975753 0.13911880 0.24911603 0.26624269 0.28022831 0.30366942
#>  [4,] 0.062564991 0.42959294 0.34084186 0.14721363 0.15578757 0.24077302
#>  [5,] 0.033461337 0.15500172 0.17389664 0.24107126 0.22156453 0.28510205
#>  [6,] 0.043706650 0.18280212 0.20209949 0.22802520 0.30803558 0.31315898
#>  [7,] 0.022349993 0.37467447 0.48058511 0.49297627 0.47128591 0.36539073
#>  [8,] 0.028725835 0.40868422 0.40580621 0.33949225 0.27970458 0.15098196
#>  [9,] 0.058062539 0.37245181 0.54770241 0.52981294 0.54388024 0.56701100
#> [10,] 0.021280707 0.12467183 0.21304309 0.25656230 0.23477360 0.14433806
#> [11,] 0.025526976 0.22570187 0.26333051 0.29225593 0.31630853 0.39334531
#> [12,] 0.029521921 0.21041646 0.28879717 0.31794143 0.29513416 0.23521219
#> [13,] 0.062248299 0.33642799 0.22520769 0.16071734 0.40122485 0.55043821
#> [14,] 0.028822222 0.11164229 0.17071265 0.30137719 0.24769500 0.20635636
#> [15,] 0.034030950 0.23389595 0.23997405 0.11176308 0.20421959 0.47965583
#> [16,] 0.022114657 0.10493426 0.08652060 0.15509455 0.15820168 0.14000475
#> [17,] 0.011320717 0.16890811 0.11285866 0.04529319 0.04025522 0.05704046
#> [18,] 0.020499209 0.11337686 0.13996588 0.07601843 0.06392082 0.13747704
#> [19,] 0.026293955 0.12132801 0.16505989 0.18754646 0.16283539 0.17059079
#> [20,] 0.010770502 0.09812854 0.09009974 0.09165222 0.12877657 0.13824599
#> [21,] 0.005636797 0.07741045 0.06541796 0.05441444 0.11304214 0.13696968
#> [22,] 0.020465043 0.38438462 0.29854159 0.09400215 0.22981913 0.52257171
#> [23,] 0.018192177 0.11279503 0.09764336 0.07162046 0.16859918 0.32200806
#> [24,] 0.026247652 0.18603044 0.25043771 0.20201187 0.17598916 0.21433715
#> [25,] 0.009693430 0.19773957 0.20864823 0.26691632 0.44907380 0.38638211
#>              V15       V16       V17       V18       V19           V2       V20
#>  [1,] 0.18360192 0.2063519 0.1819325 0.3033225 0.5608505 0.0252515968 0.6785532
#>  [2,] 0.69102237 0.7585784 0.7923616 0.8548508 0.9205961 0.0645588389 0.9159361
#>  [3,] 0.38699157 0.3275020 0.4051522 0.2051466 0.3331176 0.0862288991 0.3495480
#>  [4,] 0.33562780 0.3397874 0.1912725 0.2132349 0.3742475 0.0779440284 0.7827090
#>  [5,] 0.36341989 0.5397693 0.7210478 0.8416414 0.9471131 0.0558233811 0.9125028
#>  [6,] 0.31530120 0.3574081 0.3511453 0.2167809 0.2932186 0.0715400517 0.4067557
#>  [7,] 0.28247077 0.3025731 0.3396970 0.3406370 0.4727869 0.0269899764 0.6375364
#>  [8,] 0.15830803 0.1902422 0.1962707 0.2605856 0.2432653 0.0391126488 0.2016246
#>  [9,] 0.66911125 0.8488795 0.9594476 0.9355238 0.8059327 0.0911813724 0.7470206
#> [10,] 0.08434954 0.1739578 0.2779498 0.3649154 0.4628309 0.0369765325 0.5677790
#> [11,] 0.45660294 0.5189429 0.6152498 0.6980599 0.7596063 0.0239880136 0.8098248
#> [12,] 0.18533934 0.1380438 0.1929021 0.2213237 0.2554950 0.0412289797 0.2786055
#> [13,] 0.23650004 0.4140943 0.5708647 0.5639671 0.7851565 0.0615037326 0.8618858
#> [14,] 0.12234337 0.1601509 0.2504190 0.4861290 0.4481908 0.0479897122 0.6742983
#> [15,] 0.65251807 0.7341602 0.7924626 0.6618762 0.4736417 0.0657290051 0.3381595
#> [16,] 0.16395505 0.1657469 0.1750327 0.2387957 0.2724414 0.0460904854 0.2563290
#> [17,] 0.09847567 0.2283680 0.2940977 0.2871163 0.3959077 0.0005376669 0.5791294
#> [18,] 0.10552890 0.1140567 0.2065159 0.2375671 0.2628415 0.0103673240 0.3743388
#> [19,] 0.23019754 0.4761508 0.6631016 0.5463177 0.3189657 0.0227757393 0.3477696
#> [20,] 0.14846606 0.1818162 0.2119451 0.2798054 0.3359483 0.0128011071 0.3511398
#> [21,] 0.15020652 0.1548470 0.1312270 0.1677636 0.2859860 0.0152755567 0.3888136
#> [22,] 0.65218462 0.7624552 0.8199699 0.7346222 0.4982620 0.0354483509 0.3415624
#> [23,] 0.52971032 0.7243243 0.7817956 0.7763689 0.8282743 0.0142487624 0.8563592
#> [24,] 0.24808786 0.2391741 0.2529252 0.1114438 0.4432426 0.0137916061 0.6585831
#> [25,] 0.40032206 0.5844683 0.3995140 0.1597977 0.2974415 0.0185963101 0.4054790
#>             V21       V22        V23        V24       V25       V26       V27
#>  [1,] 0.8081464 0.8057929 0.76377113 0.88423117 0.9062519 1.0054363 1.0356358
#>  [2,] 0.9111876 0.8478570 0.86795427 0.83740372 0.6090832 0.5251241 0.4358822
#>  [3,] 0.3388377 0.3070543 0.37304580 0.28144547 0.4230373 0.4671151 0.5021273
#>  [4,] 0.8479763 0.4762907 0.26508518 0.24359305 0.2876735 0.4818491 0.4580591
#>  [5,] 0.9148812 1.0495363 0.99348125 0.76896328 0.5419175 0.5110510 0.5672835
#>  [6,] 0.5035110 0.5541703 0.54488102 0.56381322 0.5379746 0.5724828 0.6421563
#>  [7,] 0.7738260 0.8660115 0.85026395 0.85522740 0.8750308 0.9054456 0.9140618
#>  [8,] 0.2978459 0.6183066 0.78139811 0.80750110 0.8478354 0.8826389 0.9443780
#>  [9,] 0.6719109 0.3720178 0.06532297 0.04587261 0.2690180 0.3171863 0.2693559
#> [10,] 0.6454035 0.7272958 0.75470252 0.71060700 0.7584396 0.8683143 0.9655698
#> [11,] 0.8704209 0.9209385 0.93675888 0.97175827 0.9525358 0.9366450 0.9121987
#> [12,] 0.3999863 0.4770682 0.56638196 0.67005643 0.7190155 0.7618082 0.8822150
#> [13,] 0.5431226 0.6453579 0.82786910 0.67534174 0.7652302 0.8601615 0.8133428
#> [14,] 0.7643911 0.7468770 0.78698090 0.78575984 0.7831167 0.6523921 0.5681983
#> [15,] 0.5279888 0.7442992 0.84002387 0.95737925 0.8376211 0.6620260 0.6532986
#> [16,] 0.2176243 0.2044441 0.26316803 0.28319683 0.2874635 0.3395014 0.3263316
#> [17,] 0.6660960 0.8888653 1.03751009 0.87591288 0.8018153 0.8901259 0.9186149
#> [18,] 0.4990896 0.6421763 0.80610094 0.84565185 0.8661438 0.9355497 0.9147026
#> [19,] 0.4129391 0.4958963 0.55591096 0.68570877 0.8338673 0.8739024 0.8391773
#> [20,] 0.3781949 0.4296372 0.46129208 0.55040363 0.5813051 0.6298222 0.7179229
#> [21,] 0.4640740 0.5406296 0.58303597 0.68875937 0.8486730 0.8490947 0.8448291
#> [22,] 0.4292425 0.6303395 0.79586997 0.95600704 0.8203525 0.7275101 0.7831527
#> [23,] 0.9419313 0.9430263 0.92527864 0.94799744 0.9188986 0.9161805 0.8836950
#> [24,] 0.4470196 0.4080082 0.96460887 0.90727619 0.4969795 0.5793718 0.6765489
#> [25,] 0.3896453 0.4218269 0.54026546 0.47115683 0.6115730 0.8696346 0.8127576
#>             V28        V29           V3       V30       V31       V32       V33
#>  [1,] 0.9994671 0.88933997  0.040123180 0.6287198 0.4118900 0.2759704 0.2541355
#>  [2,] 0.3147884 0.31850821  0.062834070 0.2784542 0.1742796 0.1060292 0.2135527
#>  [3,] 0.7458268 0.76851274  0.058427637 0.6751261 0.7652266 0.6652660 0.4641782
#>  [4,] 0.6802791 0.64255914  0.055405979 0.6919297 0.3371924 0.2711551 0.4536180
#>  [5,] 0.5089146 0.47132045  0.057351221 0.4279328 0.2392948 0.2067634 0.3613903
#>  [6,] 0.7848179 0.83168748  0.086978066 0.8323056 0.8281368 0.7570476 0.5809902
#>  [7,] 0.8168050 0.62616871  0.033212922 0.4491294 0.3796914 0.4746547 0.4627621
#>  [8,] 0.9889280 0.77840189  0.050332529 0.5516607 0.6625926 0.6406027 0.4749195
#>  [9,] 0.2320868 0.08093118  0.138943221 0.2410228 0.2239137 0.2159364 0.2579111
#> [10,] 0.9485014 0.77738449  0.046770581 0.5962708 0.2890385 0.2400642 0.2894626
#> [11,] 0.8140661 0.68655967  0.036790618 0.5565011 0.4025320 0.2873310 0.1963349
#> [12,] 0.9623616 0.88600776  0.053404314 0.7108620 0.6526152 0.5288028 0.4059101
#> [13,] 0.6896122 0.37056373  0.071234159 0.6462157 0.7113614 0.8173896 0.6776714
#> [14,] 0.5424918 0.53469301  0.058051988 0.5468608 0.6000039 0.5657232 0.6033967
#> [15,] 0.5421013 0.42539782  0.094255604 0.4777730 0.3961634 0.1752784 0.2062297
#> [16,] 0.4456903 0.57405018  0.037480171 0.6772614 0.7290742 0.7280665 0.7316131
#> [17,] 0.9177455 0.83376245 -0.004574697 0.7197655 0.5989493 0.5210609 0.5223389
#> [18,] 0.8554408 0.83413242  0.014628254 0.7195331 0.6192996 0.5591414 0.5624715
#> [19,] 0.8468889 0.77893550  0.028631752 0.7457464 0.7998125 0.7316268 0.6552084
#> [20,] 0.8633155 0.97703990  0.015978384 0.9379666 0.8390833 0.7502133 0.7319123
#> [21,] 0.7515773 0.56412010  0.001025345 0.4074467 0.3367852 0.2712096 0.2588785
#> [22,] 0.6145996 0.42176340  0.071336014 0.4797934 0.5293862 0.3196033 0.2115053
#> [23,] 0.6916356 0.46432783  0.024606912 0.4031860 0.4196740 0.3849398 0.3456093
#> [24,] 0.7153495 0.53520348  0.062318144 0.3540392 0.2525459 0.1630142 0.1155612
#> [25,] 0.7824502 0.58047436  0.027773749 0.5462934 0.5956275 0.5355941 0.3290071
#>              V34        V35        V36        V37        V38        V39
#>  [1,] 0.21235719 0.13585411 0.01121908 0.03640674 0.09759100 0.09499694
#>  [2,] 0.30561528 0.33046653 0.32446081 0.26705945 0.26843729 0.29740690
#>  [3,] 0.57612662 0.64297739 0.38020896 0.10096184 0.30282304 0.28136774
#>  [4,] 0.66517441 0.82458443 0.84612509 0.82688581 0.85842684 0.73088732
#>  [5,] 0.37928067 0.29645873 0.19815232 0.30538488 0.41013248 0.40691328
#>  [6,] 0.45011663 0.41052910 0.26830957 0.23009740 0.31948098 0.31550245
#>  [7,] 0.35934506 0.17622176 0.27877857 0.33477552 0.36567721 0.50705047
#>  [8,] 0.21008438 0.09706409 0.03327636 0.17143955 0.20295162 0.19013129
#>  [9,] 0.21462077 0.28267679 0.12435799 0.11982146 0.19895985 0.17974553
#> [10,] 0.50491972 0.73331684 0.83811512 0.75381018 0.62543375 0.48492865
#> [11,] 0.11696764 0.13813605 0.15272614 0.20933371 0.25075843 0.24298522
#> [12,] 0.34173907 0.18781543 0.09800626 0.23412130 0.18100275 0.19328329
#> [13,] 0.53256979 0.30132984 0.08916814 0.10717844 0.39627772 0.60306010
#> [14,] 0.56164223 0.57333722 0.60909969 0.52304736 0.43416345 0.26970510
#> [15,] 0.09668088 0.39814111 0.68914079 0.71315391 0.38112111 0.09387297
#> [16,] 0.74991974 0.81488207 0.79598720 0.71969316 0.68138226 0.65236467
#> [17,] 0.59068904 0.63574648 0.56102321 0.52665021 0.34864295 0.17923217
#> [18,] 0.64982946 0.66807910 0.58766808 0.46448551 0.37333604 0.24375375
#> [19,] 0.68736110 0.70780278 0.69830336 0.67284139 0.55293587 0.35409513
#> [20,] 0.64462128 0.57757595 0.52936904 0.44237692 0.32110424 0.21913413
#> [21,] 0.25172962 0.22990360 0.25224513 0.29623968 0.28476201 0.30274063
#> [22,] 0.20219422 0.56961415 0.67975713 0.59713761 0.31934305 0.27980037
#> [23,] 0.28764710 0.20253551 0.11525702 0.20088691 0.17031529 0.13096119
#> [24,] 0.33360134 0.29500015 0.30923730 0.26845436 0.23600293 0.28122757
#> [25,] 0.33070447 0.69467271 0.71332354 0.31546448 0.07185455 0.33844430
#>                 V4        V40        V41        V42        V43        V44
#>  [1,]  0.051639214 0.11023146 0.03993122 0.11583150 0.22362894 0.24520560
#>  [2,]  0.077079712 0.29115538 0.31059507 0.29394588 0.26112445 0.23327294
#>  [3,]  0.061568713 0.10473968 0.50033141 0.78954308 0.74883383 0.55376514
#>  [4,]  0.073700902 0.49174739 0.31520877 0.25272747 0.25535847 0.32969599
#>  [5,]  0.021586579 0.43249175 0.34701946 0.37300683 0.33633302 0.26595215
#>  [6,]  0.077353079 0.26808792 0.56931761 0.63170547 0.45629495 0.43771393
#>  [7,]  0.032183084 0.45835995 0.25277897 0.20479188 0.32118663 0.36418278
#>  [8,]  0.087768794 0.11264275 0.15910435 0.06401445 0.06920908 0.14059364
#>  [9,]  0.122093756 0.04517610 0.11917344 0.16399131 0.14728182 0.14920648
#> [10,]  0.058563232 0.35175366 0.34985293 0.35002042 0.38566562 0.34736070
#> [11,]  0.061726146 0.25576636 0.25793657 0.25797481 0.20275881 0.16039442
#> [12,]  0.056001846 0.30116633 0.31402205 0.33576707 0.24480171 0.11133621
#> [13,]  0.105498240 0.37541044 0.41364719 0.49373188 0.45824440 0.41219075
#> [14,]  0.072362522 0.38135579 0.37242778 0.40149807 0.39415261 0.27195101
#> [15,]  0.082470940 0.29801211 0.42752468 0.39609098 0.24486027 0.19883748
#> [16,]  0.017293392 0.59781926 0.51638424 0.47825516 0.31513167 0.26075965
#> [17,]  0.027546203 0.05794988 0.08129480 0.01823636 0.07682023 0.02570981
#> [18,]  0.021601085 0.22413293 0.21448976 0.14487289 0.15975849 0.11380966
#> [19,]  0.039031637 0.23237906 0.23463675 0.22406844 0.20633786 0.18980389
#> [20,]  0.017195843 0.18607603 0.16729927 0.18341286 0.22507181 0.23147025
#> [21,] -0.001502924 0.24136864 0.21352840 0.19749244 0.16538231 0.16522115
#> [22,]  0.051872156 0.43675738 0.37210408 0.32407737 0.24952762 0.17873797
#> [23,]  0.033953843 0.14418561 0.09892511 0.05483070 0.07240911 0.11337481
#> [24,]  0.048088952 0.13997395 0.02528178 0.03802444 0.09158139 0.10682140
#> [25,]  0.039240433 0.56696223 0.32293654 0.18852297 0.19656094 0.27113072
#>               V45        V46        V47         V48         V49          V5
#>  [1,]  0.24671600 0.24411276 0.19540505 0.125082543 0.061419552 0.089929644
#>  [2,]  0.14107102 0.04613197 0.05153977 0.037563773 0.031986406 0.067251826
#>  [3,]  0.58557018 0.72689686 0.54945906 0.331310102 0.160479614 0.063500886
#>  [4,]  0.43852795 0.34445407 0.19185191 0.165133213 0.083225131 0.058737662
#>  [5,]  0.12515898 0.09564843 0.06176597 0.021499871 0.034299743 0.073550802
#>  [6,]  0.57789234 0.52781449 0.32890648 0.180689349 0.100562478 0.088441114
#>  [7,]  0.38899217 0.25229102 0.14603888 0.099310142 0.059839107 0.049048929
#>  [8,]  0.07584739 0.06094750 0.07061480 0.108368032 0.074907815 0.105296567
#>  [9,]  0.07205050 0.01879051 0.07364239 0.073304105 0.015862849 0.125781648
#> [10,]  0.33351549 0.24114917 0.11119239 0.048860486 0.046409634 0.046580679
#> [11,]  0.12573466 0.10188874 0.10734639 0.080945168 0.049077228 0.114145520
#> [12,]  0.12840236 0.17239125 0.19285836 0.161081856 0.075466638 0.077680598
#> [13,]  0.53364647 0.54335702 0.31130138 0.243295599 0.129428201 0.234506384
#> [14,]  0.16608803 0.13597000 0.06889413 0.046832715 0.028723028 0.110689288
#> [15,]  0.22175487 0.18180022 0.17889139 0.167372760 0.116438050 0.063010782
#> [16,]  0.25355788 0.21838605 0.19066866 0.135414492 0.072927300 0.028170695
#> [17,] -0.01567394 0.02019104 0.02795544 0.036858627 0.017846203 0.061515189
#> [18,]  0.06425919 0.04676439 0.03973643 0.041181356 0.015382642 0.053015170
#> [19,]  0.15964627 0.15259033 0.09054792 0.048488220 0.031585364 0.056052008
#> [20,]  0.19465158 0.12911724 0.07633974 0.069881474 0.039015271 0.023765134
#> [21,]  0.13305233 0.10153162 0.05336579 0.038101894 0.027109949 0.009081398
#> [22,]  0.18622525 0.15879307 0.13799663 0.119428031 0.071822361 0.046227879
#> [23,]  0.10154949 0.04909338 0.05469867 0.046486148 0.027471867 0.071278372
#> [24,]  0.07573571 0.03592211 0.02526461 0.008522495 0.011380351 0.072347449
#> [25,]  0.16646568 0.13101795 0.04116075 0.040981674 0.009337632 0.129944096
#>               V50           V51         V52         V53           V54
#>  [1,] 0.016747162  0.0210891030 0.011877820 0.002715319  0.0040796869
#>  [2,] 0.010066971  0.0220076437 0.012041206 0.012204196  0.0174802944
#>  [3,] 0.048131571  0.1015188792 0.070956933 0.031290389  0.0314333925
#>  [4,] 0.030194333  0.0227934166 0.019298414 0.013825638  0.0147234912
#>  [5,] 0.016760154  0.0070054650 0.010369914 0.014785320  0.0144498088
#>  [6,] 0.040115094  0.0358427514 0.030018788 0.018271677  0.0159876074
#>  [7,] 0.020067202  0.0177146276 0.008894556 0.008095149  0.0117465847
#>  [8,] 0.027540683  0.0130752431 0.009002473 0.003003885  0.0046177422
#>  [9,] 0.008140290  0.0264404278 0.011450362 0.025706381  0.0105531907
#> [10,] 0.020139955  0.0069668617 0.012264021 0.009004452  0.0039501180
#> [11,] 0.023338977  0.0174164467 0.014995743 0.014889283  0.0153655819
#> [12,] 0.011106391  0.0237363289 0.017533692 0.009329820  0.0107836976
#> [13,] 0.079014929  0.0238901151 0.020269957 0.033220761  0.0223564736
#> [14,] 0.016925359  0.0097457307 0.008804586 0.012363315  0.0088820837
#> [15,] 0.028877348  0.0183563258 0.012475291 0.007187538  0.0124970090
#> [16,] 0.026091619  0.0196855523 0.009176690 0.010800777  0.0092295347
#> [17,] 0.007484872 -0.0008019443 0.003138669 0.009291525  0.0090443195
#> [18,] 0.008430760  0.0074922065 0.004853333 0.009456592  0.0086798884
#> [19,] 0.020028123  0.0065377550 0.007608324 0.004106437  0.0054310828
#> [20,] 0.014769846  0.0090727579 0.008504067 0.006093535  0.0073863950
#> [21,] 0.010241292  0.0033114345 0.004316924 0.006539875  0.0115696626
#> [22,] 0.020907114  0.0182136165 0.011780456 0.004006280  0.0123957390
#> [23,] 0.008566225  0.0165390188 0.009742077 0.008732682  0.0078014028
#> [24,] 0.009541162  0.0284260166 0.039156618 0.015881138 -0.0006722203
#> [25,] 0.006815855  0.0160570990 0.007910503 0.008820241  0.0172466137
#>               V55         V56         V57          V58         V59          V6
#>  [1,] 0.002160170 0.002227480 0.003315375 0.0020042902 0.002125698  0.09065304
#>  [2,] 0.013148580 0.012808352 0.009958872 0.0117035517 0.019449551  0.12170087
#>  [3,] 0.024628613 0.008981142 0.017776792 0.0216792293 0.022382399  0.13253160
#>  [4,] 0.013273864 0.014482694 0.012753839 0.0112497018 0.010208028  0.08860317
#>  [5,] 0.001045630 0.009624588 0.002471642 0.0016580866 0.006630315  0.14102175
#>  [6,] 0.012524673 0.010437618 0.011539626 0.0111944896 0.012337176  0.10626613
#>  [7,] 0.012159517 0.008667828 0.007285817 0.0085722542 0.008755974  0.08338780
#>  [8,] 0.003807002 0.008947332 0.005143733 0.0056624654 0.006168880  0.11889339
#>  [9,] 0.014145533 0.018798724 0.019742118 0.0109652641 0.011513901  0.11883546
#> [10,] 0.007013287 0.005607940 0.003192986 0.0034433446 0.005167427  0.06782643
#> [11,] 0.006313934 0.010825082 0.008288332 0.0099612287 0.007040338  0.14816554
#> [12,] 0.007101442 0.007546709 0.005877457 0.0087589945 0.005785852  0.12583570
#> [13,] 0.041693668 0.036380143 0.033018111 0.0407313555 0.022733094  0.12529023
#> [14,] 0.009002283 0.008032092 0.007625659 0.0068905346 0.003402184  0.17203368
#> [15,] 0.006678005 0.005750470 0.011245393 0.0083088841 0.009266814  0.07488647
#> [16,] 0.011148612 0.008604304 0.005080043 0.0105346014 0.006515001  0.06274546
#> [17,] 0.009170747 0.002036786 0.003072631 0.0002575029 0.001282989  0.10894745
#> [18,] 0.006056943 0.007705121 0.005454691 0.0042452894 0.003412395  0.09117817
#> [19,] 0.010044479 0.004154414 0.012248880 0.0100995643 0.009238211  0.09573863
#> [20,] 0.005491331 0.003978755 0.003785961 0.0033242919 0.007487592  0.06355658
#> [21,] 0.006808481 0.003975666 0.003046722 0.0068569506 0.003387214 -0.01080299
#> [22,] 0.011084717 0.002777251 0.010554854 0.0058533161 0.010967062  0.13220991
#> [23,] 0.004865084 0.006779214 0.004539048 0.0041498700 0.005335138  0.11340143
#> [24,] 0.010328428 0.009187738 0.008447294 0.0171421835 0.010467348  0.15304330
#> [25,] 0.006752674 0.014616671 0.013060430 0.0044590142 0.016034626  0.20592749
#>               V60         V7         V8         V9
#>  [1,] 0.004882462 0.09583402 0.13489524 0.19503996
#>  [2,] 0.015016697 0.17441595 0.15306224 0.13054429
#>  [3,] 0.010361307 0.09831479 0.18016197 0.13341093
#>  [4,] 0.010052675 0.16392499 0.26513361 0.42324718
#>  [5,] 0.010109539 0.09704051 0.08476267 0.12323099
#>  [6,] 0.008429105 0.08778811 0.12422617 0.21295638
#>  [7,] 0.005897286 0.10518364 0.15594449 0.28534333
#>  [8,] 0.010065617 0.14043377 0.23008501 0.31399900
#>  [9,] 0.013511349 0.10590700 0.10221777 0.22598572
#> [10,] 0.004677790 0.06168248 0.07682393 0.11140562
#> [11,] 0.005698683 0.15478804 0.15541097 0.20032214
#> [12,] 0.003862894 0.12091053 0.10337484 0.15743383
#> [13,] 0.009164526 0.17472278 0.10251507 0.24592977
#> [14,] 0.010830767 0.16622798 0.15688917 0.11288512
#> [15,] 0.005595649 0.16833733 0.20868999 0.21686955
#> [16,] 0.004270244 0.09547549 0.09543439 0.05937260
#> [17,] 0.002801883 0.11455775 0.08763961 0.15487822
#> [18,] 0.004300220 0.09521292 0.07827569 0.12542486
#> [19,] 0.008380644 0.12789713 0.13428391 0.08885568
#> [20,] 0.004210000 0.06924434 0.05216983 0.08276797
#> [21,] 0.003855450 0.01771570 0.04831499 0.05979225
#> [22,] 0.008638033 0.11300161 0.20200984 0.30263557
#> [23,] 0.003921870 0.12295385 0.08058569 0.11723889
#> [24,] 0.006193105 0.13599609 0.19497040 0.15549141
#> [25,] 0.007726371 0.18742883 0.15260868 0.14739350
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551