Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11          V12        V13        V14
#>  [1,] 0.082481206  0.56790437 0.44753080  0.190508479 0.23316299 0.32668028
#>  [2,] 0.054778024  0.22081055 0.23998688  0.281807940 0.28769964 0.25697557
#>  [3,] 0.014543933  0.20263284 0.31342825  0.323852559 0.27175604 0.24686346
#>  [4,] 0.019190409  0.31962965 0.43230361  0.442515347 0.35052702 0.26912223
#>  [5,] 0.056986051 -0.04315297 0.14378343  0.260004405 0.31569142 0.36683191
#>  [6,] 0.058621441  0.40732621 0.57312692  0.509711856 0.52762589 0.52459700
#>  [7,] 0.028511465  0.18670479 0.12630868  0.026487597 0.09101807 0.30393152
#>  [8,] 0.041192734  0.38537470 0.41752659  0.441599978 0.48879362 0.44677572
#>  [9,] 0.027458331  0.24514185 0.29456612  0.341862936 0.31259152 0.29512573
#> [10,] 0.053099709  0.20537167 0.29544649  0.406895681 0.50333969 0.45712916
#> [11,] 0.040576746  0.31024742 0.33790629  0.316728924 0.23877625 0.10653030
#> [12,] 0.021036539  0.49107551 0.57768397  0.549902902 0.50821336 0.36027702
#> [13,] 0.032037925  0.28126968 0.40046441  0.521718054 0.50689987 0.49965146
#> [14,] 0.029831048  0.22693009 0.18517666  0.216725070 0.23304248 0.14361069
#> [15,] 0.014815292  0.12363315 0.13428049  0.103823790 0.07350290 0.12111232
#> [16,] 0.015548028  0.12272796 0.13035331  0.111236512 0.15037468 0.14150714
#> [17,] 0.021996980  0.12715780 0.15956039  0.200473280 0.19213064 0.17568533
#> [18,] 0.015242647  0.15866308 0.18573295  0.166250080 0.25926445 0.35499636
#> [19,] 0.008855895  0.14017433 0.19453691  0.136125377 0.10732068 0.14049781
#> [20,] 0.029451715  0.07246066 0.10623140  0.180502749 0.20941019 0.11466232
#> [21,] 0.005960907  0.26945410 0.28418388  0.265039519 0.31141329 0.43954271
#> [22,] 0.008006937  0.12751294 0.11663263 -0.002436952 0.04268139 0.05824561
#> [23,] 0.020252479  0.09347238 0.11292672  0.162667036 0.20587419 0.22601991
#> [24,] 0.015218574  0.08709568 0.07606943  0.061144546 0.08065207 0.10700509
#> [25,] 0.003129006  0.17691955 0.19076334  0.103104553 0.16399978 0.06463391
#>              V15        V16        V17       V18       V19           V2
#>  [1,] 0.52329155 0.54938437 0.20157778 0.1643175 0.3614443  0.087224555
#>  [2,] 0.25405602 0.26610398 0.28075893 0.1917253 0.2713938  0.062839210
#>  [3,] 0.21711753 0.20232156 0.23450752 0.2680331 0.3317467  0.029461038
#>  [4,] 0.26093612 0.36987289 0.39101453 0.3811994 0.4210821  0.025497303
#>  [5,] 0.52613369 0.54693658 0.77034006 0.9034206 0.9970915  0.047123039
#>  [6,] 0.64650290 0.80085466 0.93398126 0.9507917 0.8217515  0.090601379
#>  [7,] 0.52887857 0.65672737 0.76351079 0.6725032 0.7314992  0.032009917
#>  [8,] 0.28720680 0.30436998 0.37853278 0.3967077 0.5617319  0.032784890
#>  [9,] 0.30001456 0.35011530 0.39140482 0.4956160 0.6058990  0.020787526
#> [10,] 0.56860653 0.68802422 0.88908088 0.9322853 0.9066870  0.113555910
#> [11,] 0.11748248 0.18119185 0.18509949 0.2428998 0.2368226  0.035906470
#> [12,] 0.25351344 0.30563067 0.29813996 0.3261653 0.5407392  0.039011750
#> [13,] 0.57542421 0.74600244 0.83546472 0.9379248 1.0224258  0.052809820
#> [14,] 0.28852318 0.28535278 0.21952945 0.2111508 0.2363730  0.017831694
#> [15,] 0.09129475 0.10471044 0.21184957 0.2901369 0.3182564  0.011441669
#> [16,] 0.11335843 0.10575648 0.10492586 0.1759378 0.3349002  0.012872140
#> [17,] 0.21858183 0.50180271 0.70441384 0.6173796 0.3700642  0.017522532
#> [18,] 0.42604690 0.46642413 0.47062602 0.4904926 0.5257953  0.015608418
#> [19,] 0.16264055 0.17282253 0.19221365 0.1100453 0.3228233 -0.007413725
#> [20,] 0.09884448 0.18977516 0.23477353 0.3763950 0.2894497  0.009699102
#> [21,] 0.62287152 0.79846886 0.92267498 0.9117607 0.8349386  0.014426789
#> [22,] 0.12627938 0.05032142 0.08894456 0.1916602 0.3234490  0.017707869
#> [23,] 0.21367640 0.27336325 0.32318669 0.2494817 0.2996155  0.038792185
#> [24,] 0.13569670 0.22421224 0.26586123 0.3640708 0.4154089  0.017765112
#> [25,] 0.21417138 0.53041205 0.75761846 0.8065633 0.6206691  0.006530713
#>             V20       V21       V22         V23         V24       V25       V26
#>  [1,] 0.8761109 0.9076100 0.3700147  0.14783942  0.15309205 0.1966575 0.3342260
#>  [2,] 0.3017865 0.3000985 0.3246373  0.42860574  0.45744583 0.5312302 0.5809128
#>  [3,] 0.3509242 0.4028963 0.4471459  0.44526773  0.55572134 0.6175217 0.7057554
#>  [4,] 0.5219315 0.7190249 0.9074174  0.96415120  0.95424482 0.9256750 0.8334741
#>  [5,] 1.0128123 0.9302852 0.9316293  0.97314624  0.91018890 0.7008634 0.5703306
#>  [6,] 0.7549081 0.7142451 0.3519681 -0.04196417 -0.05567352 0.2552530 0.3326008
#>  [7,] 0.8556750 0.9738402 1.0273329  0.83972668  0.49727940 0.2830783 0.4007797
#>  [8,] 0.7138220 0.6956353 0.8022372  0.85156498  0.84204696 0.9044658 0.9497395
#>  [9,] 0.7218617 0.7958971 0.8214511  0.83928285  0.92389374 0.9651386 0.9491859
#> [10,] 0.8046353 0.7881928 0.5776980  0.35851476  0.36519836 0.2445771 0.1127498
#> [11,] 0.2418304 0.3748844 0.6175327  0.75529952  0.81910879 0.9248256 1.0100973
#> [12,] 0.7437201 0.8683790 0.8132087  0.69585158  0.72356501 0.7386394 0.8401095
#> [13,] 1.0238713 0.8387464 0.6427453  0.57943629  0.46412563 0.2924145 0.1181681
#> [14,] 0.2887502 0.2606771 0.4103093  0.42998431  0.27190369 0.1769294 0.4165762
#> [15,] 0.5098664 0.6406624 0.7618217  0.89683615  0.89090758 0.9249988 0.9785371
#> [16,] 0.3289116 0.2529001 0.3228674  0.48147478  0.63773221 0.6301950 0.6235701
#> [17,] 0.3188853 0.4210582 0.5316823  0.59079920  0.66366640 0.8266879 0.9468794
#> [18,] 0.6076414 0.6211637 0.6378840  0.69348486  0.75179939 0.7145971 0.7796330
#> [19,] 0.4975025 0.3811700 0.4250390  0.90725849  0.86022973 0.5874635 0.6536357
#> [20,] 0.2865637 0.6199825 0.6769698  0.70128887  0.81898156 0.7670182 0.6603970
#> [21,] 0.7404032 0.7349505 0.7911044  0.64508524  0.54629768 0.6487407 0.7140459
#> [22,] 0.3504688 0.4576844 0.8825907  0.96242085  0.87131659 0.6603353 0.6815464
#> [23,] 0.3521478 0.3528844 0.3706339  0.38721477  0.52660510 0.6517884 0.5927163
#> [24,] 0.4179183 0.4187195 0.3580292  0.29928677  0.37997443 0.3393402 0.4143560
#> [25,] 0.3392306 0.3783125 0.2586137  0.37090179  0.75256992 0.8900040 0.8296302
#>              V27         V28       V29          V3       V30       V31
#>  [1,] 0.25140798 0.535210907 0.5272016 0.052302141 0.6821710 0.3203517
#>  [2,] 0.66800610 0.806678628 0.8060372 0.052787705 0.8121748 0.8328520
#>  [3,] 0.88442930 0.978947517 0.8958555 0.039565079 0.6936225 0.6383354
#>  [4,] 0.71834187 0.575798728 0.3640657 0.031556455 0.3092445 0.2851240
#>  [5,] 0.50253285 0.500683879 0.4964025 0.086895173 0.4020500 0.2474627
#>  [6,] 0.28066039 0.247952122 0.0105177 0.146425165 0.2300241 0.1916644
#>  [7,] 0.51420875 0.383750388 0.3290865 0.009410223 0.5409505 0.2870048
#>  [8,] 0.94911032 0.748916307 0.4451486 0.032659581 0.4233216 0.4352283
#>  [9,] 0.92292171 0.860123995 0.7870724 0.029553460 0.6162487 0.3982589
#> [10,] 0.11490024 0.144576705 0.1490715 0.116055695 0.1496478 0.2621628
#> [11,] 0.98406162 0.946959674 0.8396634 0.026250930 0.7253557 0.7414733
#> [12,] 0.92840595 0.986101034 0.8387102 0.056499033 0.5401809 0.4108956
#> [13,] 0.06342345 0.002335003 0.1032172 0.067263127 0.2785195 0.2775793
#> [14,] 0.61084709 0.655751895 0.7047574 0.020034285 0.7887467 0.5906363
#> [15,] 0.92883405 0.827642474 0.7610166 0.013122154 0.6857530 0.6484804
#> [16,] 0.56892625 0.744694034 0.9847727 0.022165815 0.9840013 0.9181759
#> [17,] 0.95131917 0.894763330 0.8108370 0.021203214 0.8253766 0.8340873
#> [18,] 0.90785806 0.839650389 0.6577955 0.020501615 0.5830252 0.5121238
#> [19,] 0.68817560 0.722096074 0.5625398 0.019063627 0.3807689 0.2432879
#> [20,] 0.52412552 0.351512582 0.4319580 0.010019130 0.4401782 0.5065401
#> [21,] 0.67256274 0.566109930 0.5479285 0.025740651 0.6455239 0.5973092
#> [22,] 0.57513314 0.682300013 0.8234837 0.042366745 0.7794490 0.5674427
#> [23,] 0.51383043 0.618950272 0.6961461 0.030658252 0.7093654 0.6640961
#> [24,] 0.62415298 0.851534904 0.9633113 0.023552084 0.8779884 0.7334880
#> [25,] 0.78461069 0.815436179 0.5761001 0.017065231 0.3601045 0.3739145
#>             V32       V33       V34       V35        V36        V37        V38
#>  [1,] 0.3711871 0.6506747 0.8015413 0.8689107 0.85261890 0.83790766 0.88235722
#>  [2,] 0.7275709 0.5369803 0.5405638 0.5676037 0.46352415 0.36722849 0.42309548
#>  [3,] 0.5575521 0.3907078 0.2940323 0.1587536 0.13845568 0.19952460 0.11812864
#>  [4,] 0.3650852 0.4224632 0.4229269 0.2820126 0.24682580 0.29647105 0.29415825
#>  [5,] 0.1686311 0.1827177 0.2766571 0.2960703 0.26829170 0.35462174 0.46878294
#>  [6,] 0.2067402 0.2819211 0.2248216 0.3402264 0.03499801 0.05195522 0.21107277
#>  [7,] 0.4421673 0.5252276 0.2806321 0.1844230 0.30039006 0.32049430 0.31376993
#>  [8,] 0.5184899 0.4846048 0.3913740 0.1741807 0.28891653 0.37083552 0.34647813
#>  [9,] 0.2760206 0.2265745 0.1261371 0.1227503 0.10272161 0.18401153 0.25589755
#> [10,] 0.3356915 0.3689250 0.4920233 0.5624126 0.27804662 0.14502653 0.30477769
#> [11,] 0.6038703 0.4263804 0.2160321 0.1260355 0.09490592 0.13811872 0.15994814
#> [12,] 0.4006327 0.3259825 0.2858127 0.1208973 0.32237857 0.23089291 0.36424849
#> [13,] 0.2889257 0.2722304 0.4071363 0.3839526 0.30375358 0.44113399 0.41477669
#> [14,] 0.4286767 0.6992208 0.8499198 0.7932667 0.80494533 0.66884521 0.59105978
#> [15,] 0.6158421 0.5963657 0.5878476 0.5806118 0.53302538 0.46675177 0.33145809
#> [16,] 0.8443276 0.7804792 0.6730197 0.5485919 0.45709325 0.39896423 0.20336702
#> [17,] 0.7281063 0.6943710 0.7029896 0.7645295 0.76737382 0.72375771 0.63972348
#> [18,] 0.3420932 0.2821999 0.1313897 0.2597091 0.43111261 0.43309405 0.24636549
#> [19,] 0.1801003 0.1502249 0.3127942 0.2659138 0.28428590 0.23450030 0.19156243
#> [20,] 0.5796771 0.7397683 0.6423954 0.6062400 0.60016174 0.56765274 0.43228906
#> [21,] 0.4211467 0.3752731 0.3662732 0.3012810 0.28893006 0.33729043 0.27025916
#> [22,] 0.6071062 0.7963538 0.9073333 0.8863528 0.69144896 0.42991408 0.24382272
#> [23,] 0.6626761 0.6739035 0.7253591 0.7975265 0.74541820 0.57808608 0.39371813
#> [24,] 0.6010378 0.6631885 0.6000369 0.4736327 0.50313343 0.50997779 0.41742132
#> [25,] 0.4081483 0.3505445 0.1653810 0.2663400 0.32975027 0.13556249 0.08537835
#>             V39         V4        V40        V41        V42        V43
#>  [1,] 0.7124317 0.09769924 0.50134212 0.31924286 0.26536601 0.21431919
#>  [2,] 0.4205768 0.05293072 0.36711887 0.53865343 0.59348118 0.55872625
#>  [3,] 0.1492930 0.04992054 0.27163140 0.30910376 0.35538911 0.24661976
#>  [4,] 0.4456387 0.03621522 0.46151439 0.31494083 0.13888006 0.28906731
#>  [5,] 0.4629964 0.07148077 0.28807380 0.27587618 0.35925792 0.31909004
#>  [6,] 0.1696176 0.13932656 0.01289871 0.09081448 0.17583633 0.14605029
#>  [7,] 0.2119752 0.01213554 0.13529986 0.08391841 0.26284491 0.20565972
#>  [8,] 0.3408759 0.04060146 0.26669620 0.15343550 0.25440652 0.36899740
#>  [9,] 0.2666326 0.05759630 0.26432218 0.22623011 0.24695526 0.25519695
#> [10,] 0.3428885 0.14313312 0.17867262 0.07294030 0.05248093 0.08594303
#> [11,] 0.1681215 0.04706431 0.06680900 0.13373645 0.09475609 0.13202023
#> [12,] 0.6090611 0.07005604 0.44131059 0.16663455 0.40055191 0.47061782
#> [13,] 0.1657629 0.04676881 0.33983526 0.63019978 0.44188078 0.18212296
#> [14,] 0.3579350 0.03819752 0.68519401 0.74573060 0.61534363 0.37689967
#> [15,] 0.1343866 0.03140459 0.16629824 0.18053404 0.11873039 0.08980603
#> [16,] 0.1180715 0.03689334 0.16700439 0.17509145 0.17104124 0.24029660
#> [17,] 0.4358148 0.02845686 0.27591796 0.23792637 0.19935986 0.20554847
#> [18,] 0.1716954 0.01399267 0.18698527 0.25589086 0.27763249 0.20119327
#> [19,] 0.2401735 0.01159789 0.14866912 0.08652310 0.06008098 0.09271140
#> [20,] 0.4560960 0.00737188 0.67035204 0.74060884 0.67679431 0.69589726
#> [21,] 0.2160499 0.04566527 0.29909870 0.23648263 0.17622360 0.11442320
#> [22,] 0.2583866 0.03449597 0.34924193 0.30574922 0.20352457 0.30789256
#> [23,] 0.3007945 0.02924837 0.25698153 0.22732653 0.22599060 0.15055023
#> [24,] 0.2934721 0.02141147 0.18688060 0.12602901 0.16255947 0.18179859
#> [25,] 0.1465935 0.02552072 0.16361538 0.08916793 0.07342812 0.07849143
#>              V44        V45        V46        V47         V48         V49
#>  [1,] 0.25813450 0.40599503 0.29538263 0.14246733 0.125144800 0.071501995
#>  [2,] 0.49806583 0.54980980 0.54965646 0.40445464 0.256679466 0.138150265
#>  [3,] 0.08540044 0.15747639 0.17117348 0.19734310 0.167451385 0.074676045
#>  [4,] 0.27552232 0.32397401 0.28599254 0.18698235 0.101548114 0.064226689
#>  [5,] 0.26077435 0.17858746 0.07938139 0.03356780 0.020175231 0.028782403
#>  [6,] 0.13596111 0.07365013 0.01944597 0.07463640 0.071721706 0.012411489
#>  [7,] 0.14007190 0.09872835 0.00601919 0.05800196 0.081804042 0.047459354
#>  [8,] 0.43313159 0.40917137 0.28246280 0.14643151 0.090318403 0.039787833
#>  [9,] 0.20401546 0.16753855 0.14710938 0.12342455 0.082535819 0.055631033
#> [10,] 0.05175120 0.03915383 0.05525618 0.10051629 0.103172490 0.071639546
#> [11,] 0.15407593 0.10667822 0.08306555 0.10184812 0.095039171 0.059932371
#> [12,] 0.48258057 0.41192758 0.13592512 0.12859735 0.181990963 0.103822788
#> [13,] 0.14212158 0.05258907 0.05278482 0.06813245 0.022947707 0.022199038
#> [14,] 0.19872554 0.23534549 0.31103255 0.22729955 0.101551280 0.052458265
#> [15,] 0.03827345 0.03885439 0.02899820 0.02225372 0.029423492 0.010129049
#> [16,] 0.30236534 0.27458701 0.17405292 0.07876941 0.023285486 0.013463765
#> [17,] 0.15647293 0.15620485 0.15955251 0.08946619 0.047961680 0.025115895
#> [18,] 0.12094887 0.12067737 0.10157105 0.08689071 0.074930374 0.047002401
#> [19,] 0.11374238 0.07244709 0.03622495 0.01447228 0.003328498 0.011462662
#> [20,] 0.58932911 0.34734947 0.22165964 0.15354149 0.107522974 0.019468514
#> [21,] 0.13254904 0.18206255 0.13476091 0.14416688 0.104459232 0.046183978
#> [22,] 0.23573594 0.12472419 0.06185465 0.04006120 0.035486974 0.010928292
#> [23,] 0.19924959 0.14637342 0.10793371 0.07136930 0.058755232 0.023792195
#> [24,] 0.17322923 0.13244351 0.08784502 0.06220338 0.079303030 0.043145315
#> [25,] 0.08864007 0.08685604 0.08845820 0.03427512 0.035778419 0.007414139
#>                V5         V50         V51          V52          V53         V54
#>  [1,] 0.073877469 0.031165113 0.024158384  0.018357507 0.0157340639 0.028102898
#>  [2,] 0.063128782 0.046278963 0.051498993  0.037161298 0.0174439142 0.018915681
#>  [3,] 0.066354522 0.013411774 0.020515954  0.014398730 0.0121037398 0.010394633
#>  [4,] 0.045732040 0.025085793 0.013147779  0.005843351 0.0119224758 0.013905043
#>  [5,] 0.155118695 0.002304919 0.001695353  0.001434341 0.0029711815 0.027388488
#>  [6,] 0.116810137 0.010311810 0.031255325  0.014577285 0.0305954807 0.012099144
#>  [7,] 0.032937717 0.020813053 0.015745614  0.012390063 0.0003837161 0.011401297
#>  [8,] 0.083572140 0.022636711 0.010528534  0.009117195 0.0134287551 0.012324546
#>  [9,] 0.090759041 0.024303380 0.017786476  0.013100729 0.0135459472 0.007506171
#> [10,] 0.135267366 0.026051201 0.021525257  0.032169429 0.0111972671 0.012401826
#> [11,] 0.049439570 0.020093021 0.013690762  0.012413848 0.0042849641 0.005837427
#> [12,] 0.116861650 0.024373789 0.033379858  0.019072499 0.0157989663 0.014466107
#> [13,] 0.049912622 0.019020429 0.003953503  0.032702628 0.0197590113 0.013933748
#> [14,] 0.075237068 0.045134247 0.035018595  0.026438085 0.0160806361 0.015404006
#> [15,] 0.068379295 0.008023723 0.005674477  0.004875316 0.0069532951 0.005039604
#> [16,] 0.031836095 0.015911154 0.011740157  0.005951210 0.0089196889 0.009587014
#> [17,] 0.031884530 0.022655717 0.010850139  0.011431298 0.0033178462 0.005337134
#> [18,] 0.005848722 0.020181308 0.008639537  0.005478605 0.0058267306 0.009757366
#> [19,] 0.026545448 0.008780903 0.023134675  0.030659938 0.0117792990 0.000292634
#> [20,] 0.033873797 0.013842261 0.006766057 -0.003815128 0.0069524529 0.005413954
#> [21,] 0.087353521 0.018719595 0.007363785  0.008262891 0.0055744543 0.004136719
#> [22,] 0.069876978 0.016717364 0.013038023  0.014670388 0.0206356596 0.015925970
#> [23,] 0.068348740 0.010205009 0.011438550  0.005687685 0.0107176502 0.011512292
#> [24,] 0.028395778 0.012079657 0.006891666  0.009271692 0.0088583212 0.011124084
#> [25,] 0.081277882 0.017941064 0.005065564  0.023588965 0.0105036076 0.013873699
#>                V55         V56         V57         V58           V59         V6
#>  [1,]  0.021904617 0.016100980 0.012428207 0.014139707  0.0133629461 0.10297778
#>  [2,]  0.016101285 0.010175236 0.011516339 0.012772160  0.0137900038 0.11105459
#>  [3,]  0.004682047 0.004538212 0.005048861 0.008404872  0.0062214350 0.10484353
#>  [4,]  0.013646712 0.010993967 0.008032759 0.010644820  0.0085011803 0.06418686
#>  [5,]  0.021839446 0.016368881 0.013216536 0.012100601  0.0153659175 0.12374504
#>  [6,]  0.016165298 0.019210716 0.019151638 0.017110290  0.0135612944 0.08397169
#>  [7,]  0.001681772 0.012439873 0.001252784 0.009510936 -0.0009780637 0.03514642
#>  [8,]  0.011902024 0.013563819 0.010062997 0.012051568  0.0114561566 0.10775217
#>  [9,]  0.005868856 0.006328840 0.005307239 0.004507065  0.0048153218 0.12008367
#> [10,]  0.018009801 0.010507052 0.002380127 0.011040535  0.0221311972 0.02714239
#> [11,]  0.004700495 0.007410771 0.008523575 0.008987587  0.0074152684 0.10700685
#> [12,]  0.011116463 0.003381536 0.006536927 0.003005036  0.0050932859 0.17061148
#> [13,]  0.029343047 0.016526242 0.004870111 0.012768196  0.0125076270 0.05025106
#> [14,]  0.011256676 0.006471563 0.009334374 0.004573113  0.0077008597 0.11926787
#> [15,]  0.005532421 0.005760591 0.004306449 0.002909348  0.0023828848 0.11273634
#> [16,]  0.002128788 0.003365714 0.004638798 0.003730268  0.0084726362 0.02815737
#> [17,]  0.008097246 0.004662620 0.010703084 0.011356695  0.0068736377 0.07906771
#> [18,]  0.008633842 0.006293942 0.007012491 0.005209326  0.0046586197 0.03276063
#> [19,]  0.006884143 0.007044044 0.005570834 0.014354666  0.0076487161 0.10385272
#> [20,] -0.002429151 0.012127475 0.009707062 0.007082699  0.0109433455 0.07293491
#> [21,]  0.008327944 0.008071366 0.008293098 0.004040888  0.0063322519 0.12643087
#> [22,]  0.006390310 0.013678799 0.023904768 0.008886473  0.0077329070 0.06903143
#> [23,]  0.013233109 0.009577764 0.011434556 0.007807442  0.0112985410 0.10204353
#> [24,]  0.008263846 0.004729173 0.005395548 0.004704027  0.0087160304 0.09729907
#> [25,]  0.011670441 0.006070315 0.005489623 0.002258833  0.0067027936 0.06603973
#>               V60         V7         V8         V9
#>  [1,] 0.011578638 0.22478235 0.42189583 0.57982679
#>  [2,] 0.007793564 0.11300156 0.16097314 0.20057700
#>  [3,] 0.003640607 0.09911039 0.10801141 0.14830067
#>  [4,] 0.006013458 0.07132749 0.11767822 0.24115738
#>  [5,] 0.010371775 0.09043251 0.05993000 0.07528759
#>  [6,] 0.016468559 0.09151546 0.04550451 0.21152860
#>  [7,] 0.008068743 0.03996522 0.06928650 0.21581891
#>  [8,] 0.006023344 0.14578276 0.15211711 0.32447669
#>  [9,] 0.004373053 0.15062106 0.13501002 0.18301760
#> [10,] 0.011553732 0.06954072 0.04148262 0.09393816
#> [11,] 0.007302435 0.12419987 0.15950921 0.19891065
#> [12,] 0.004378179 0.18196816 0.23938515 0.32394641
#> [13,] 0.005336289 0.06047619 0.05939636 0.13861592
#> [14,] 0.007773408 0.17287776 0.14558986 0.08989909
#> [15,] 0.004416586 0.11426596 0.09938116 0.13962886
#> [16,] 0.003260827 0.07070884 0.08133389 0.08507781
#> [17,] 0.009289502 0.11816577 0.10677654 0.07267364
#> [18,] 0.002930097 0.08370056 0.12579520 0.15306548
#> [19,] 0.004702522 0.10382562 0.18207288 0.13515412
#> [20,] 0.007704183 0.09520771 0.08542201 0.06211756
#> [21,] 0.006341441 0.15721484 0.15885995 0.21832595
#> [22,] 0.013121396 0.10428155 0.07898914 0.01784026
#> [23,] 0.004292435 0.11101604 0.07154327 0.09117777
#> [24,] 0.005990130 0.07190219 0.05894815 0.10538906
#> [25,] 0.003561128 0.02560165 0.01808887 0.05816972
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841