Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.110881156 0.52981125 0.37417703 0.11514539 0.09464128 0.20873927
#>  [2,] 0.077684278 0.56914181 0.46991151 0.18123501 0.25883566 0.30815251
#>  [3,] 0.100892510 0.59860185 0.48043797 0.24206428 0.11893757 0.32869029
#>  [4,] 0.039135991 0.21011204 0.22381681 0.28355824 0.32518116 0.29001722
#>  [5,] 0.015149628 0.31640892 0.28281422 0.31166015 0.24746875 0.17773065
#>  [6,] 0.055525810 0.13533516 0.26596798 0.36350487 0.45260792 0.54255392
#>  [7,] 0.012702001 0.24995154 0.35857530 0.37380445 0.39359265 0.33373457
#>  [8,] 0.007432708 0.22466208 0.33617708 0.31366709 0.27351905 0.22862828
#>  [9,] 0.038223633 0.29875963 0.34100306 0.34127809 0.25117848 0.10812625
#> [10,] 0.038030452 0.42841478 0.54059936 0.55428464 0.66355596 0.67944952
#> [11,] 0.039175689 0.29558614 0.33197476 0.38555138 0.39876554 0.38352583
#> [12,] 0.073101068 0.16412284 0.20496961 0.24297770 0.28874174 0.29622875
#> [13,] 0.031536643 0.31259127 0.31400257 0.21274411 0.15519743 0.17749867
#> [14,] 0.029628312 0.30831528 0.37620290 0.39611223 0.37670302 0.34226750
#> [15,] 0.015309099 0.09905671 0.12018429 0.14625758 0.16088896 0.17174915
#> [16,] 0.012258225 0.08766245 0.07636330 0.05035509 0.07829957 0.08791870
#> [17,] 0.007680511 0.07902607 0.06769357 0.09770666 0.14723972 0.12264119
#> [18,] 0.071559141 0.25665422 0.25899022 0.10465304 0.11870837 0.45993986
#> [19,] 0.011808442 0.11916367 0.18635116 0.33547185 0.32575695 0.25025109
#> [20,] 0.010903414 0.12896231 0.16233540 0.08064826 0.04584179 0.16060684
#> [21,] 0.013428434 0.34177993 0.28529289 0.11675016 0.23146149 0.49852997
#> [22,] 0.016243424 0.15137938 0.12621455 0.07532964 0.06271662 0.12397746
#> [23,] 0.021124729 0.03807047 0.06154929 0.16246252 0.11296307 0.09984393
#> [24,] 0.018739520 0.09171630 0.09268082 0.02806378 0.05458859 0.10072418
#> [25,] 0.002201495 0.16091180 0.10631322 0.14447506 0.25200524 0.20741798
#>              V15        V16       V17        V18       V19          V2
#>  [1,] 0.39787403 0.39889903 0.2904082 0.23048873 0.3572981 0.108560194
#>  [2,] 0.57553890 0.57639870 0.1541817 0.09998298 0.3819293 0.092791085
#>  [3,] 0.46604603 0.46377718 0.1932702 0.16851242 0.3805748 0.112648401
#>  [4,] 0.35705913 0.45887062 0.5348746 0.58489032 0.5960627 0.055701996
#>  [5,] 0.14810210 0.18599453 0.1885969 0.21216678 0.1640434 0.024669360
#>  [6,] 0.58713952 0.61425286 0.7151112 0.80962762 0.9065334 0.056247365
#>  [7,] 0.09355112 0.12205610 0.3441114 0.55156954 0.6665148 0.012748762
#>  [8,] 0.11622976 0.06356527 0.1747624 0.32152724 0.4443294 0.027522280
#>  [9,] 0.07865431 0.12420406 0.1704024 0.22443854 0.2213881 0.038020045
#> [10,] 0.73360947 0.75370649 0.8796853 0.94477879 0.8728696 0.027239018
#> [11,] 0.29145280 0.33286115 0.3108657 0.29293176 0.3408886 0.045755421
#> [12,] 0.34499267 0.33737974 0.3801076 0.18218896 0.2936513 0.081709237
#> [13,] 0.33243798 0.47573072 0.4367851 0.44489308 0.5553023 0.057253210
#> [14,] 0.30079598 0.32919637 0.3966756 0.50759050 0.6784707 0.038242697
#> [15,] 0.20056951 0.25409248 0.3090843 0.41721937 0.4583745 0.022029225
#> [16,] 0.01115409 0.10660886 0.1575172 0.39873506 0.3352427 0.006620854
#> [17,] 0.08601067 0.09180124 0.1007399 0.15808705 0.1813630 0.007165042
#> [18,] 0.68876430 0.69560127 0.7268143 0.57426102 0.3850918 0.041413757
#> [19,] 0.22609354 0.49239836 0.7216126 0.72667901 0.5469361 0.027962899
#> [20,] 0.30564218 0.21134567 0.1292152 0.16578095 0.1833380 0.014130689
#> [21,] 0.64002102 0.77659368 0.8736361 0.80732313 0.5824276 0.017495384
#> [22,] 0.12119410 0.11484866 0.2124342 0.25004600 0.3234597 0.002321140
#> [23,] 0.08942441 0.13107618 0.1360875 0.04909505 0.1390101 0.033439452
#> [24,] 0.10694838 0.11059536 0.1464551 0.18520922 0.1913496 0.016834129
#> [25,] 0.09522806 0.21998866 0.2687385 0.36061842 0.5924177 0.025906112
#>             V20        V21        V22        V23        V24        V25
#>  [1,] 0.8378253 0.98274672 0.45433087 0.06892085 0.20943433 0.15903781
#>  [2,] 0.8863494 0.85451880 0.35646483 0.17776045 0.16924149 0.17142155
#>  [3,] 0.9007796 0.94783757 0.45463342 0.12573724 0.12441827 0.26258175
#>  [4,] 0.7002865 0.76514555 0.87484503 0.99433085 0.99911816 0.99564884
#>  [5,] 0.1149522 0.08737071 0.23347060 0.45069150 0.59548659 0.71036523
#>  [6,] 0.9320178 0.92626359 0.86324016 0.85192409 0.84522882 0.63334192
#>  [7,] 0.7246033 0.78672614 0.83727579 0.88413821 0.81964730 0.80446637
#>  [8,] 0.4792298 0.39080639 0.33002275 0.39246303 0.49782593 0.61339436
#>  [9,] 0.2531973 0.37140204 0.58597691 0.75774291 0.83991590 0.88066640
#> [10,] 0.7835619 0.61621378 0.35624132 0.08100033 0.01694481 0.24460542
#> [11,] 0.5794327 0.62660632 0.58995700 0.60872481 0.63212909 0.62739972
#> [12,] 0.3506809 0.37903150 0.38217808 0.43310674 0.39971345 0.45151345
#> [13,] 0.6715922 0.78080674 0.93346675 0.93355415 0.77986393 0.53643284
#> [14,] 0.7849195 0.84480257 0.88233316 0.90567187 0.96218044 0.94635871
#> [15,] 0.5228274 0.56595610 0.51266745 0.52265117 0.61195394 0.56266822
#> [16,] 0.2985587 0.20093921 0.19755962 0.18259433 0.19919890 0.02317465
#> [17,] 0.1835202 0.23373259 0.39598210 0.49956438 0.63281148 0.71266358
#> [18,] 0.2726041 0.53040289 0.77407114 0.86858336 0.94781907 0.92395535
#> [19,] 0.3829325 0.41244353 0.45942758 0.57667837 0.72860222 0.84848359
#> [20,] 0.2449414 0.23614399 0.51532756 0.68502716 0.64063234 0.64776482
#> [21,] 0.4128134 0.49918971 0.69002248 0.79537211 0.91104469 0.81603053
#> [22,] 0.5068545 0.63228874 0.78850545 0.93439838 0.90707404 0.89556666
#> [23,] 0.1973127 0.17350225 0.09068085 0.19159872 0.14292687 0.11434501
#> [24,] 0.3036205 0.43594210 0.56128368 0.74220950 0.81075536 0.84783515
#> [25,] 0.6918636 0.67340606 0.57035879 0.52509807 0.42791321 0.58397656
#>             V26       V27       V28       V29            V3       V30       V31
#>  [1,] 0.3065826 0.1905772 0.2830501 0.4386668  0.0812364504 0.6210286 0.2727603
#>  [2,] 0.3031499 0.2440695 0.6064039 0.5422214  0.0529105357 0.7051611 0.3250452
#>  [3,] 0.5177007 0.3283834 0.5902306 0.5848923  0.0773687952 0.8432074 0.5289858
#>  [4,] 1.0537172 1.0467898 1.0509449 0.7969042 -0.0003464701 0.5598277 0.5269899
#>  [5,] 0.8297304 0.9282401 0.9482169 0.9025871  0.0235108987 0.9290038 0.9116552
#>  [6,] 0.4607453 0.3424089 0.2637120 0.2725640  0.0704772212 0.2830193 0.2628649
#>  [7,] 0.6585000 0.3206532 0.4641294 0.6237084  0.0102767129 0.7243707 0.8025031
#>  [8,] 0.7891986 0.9736430 1.0378471 0.9566736  0.0271843228 0.6904195 0.5168960
#>  [9,] 0.9240929 0.9569685 0.9804161 0.8682503  0.0328684982 0.7170352 0.7397745
#> [10,] 0.4450944 0.4596562 0.4023781 0.2168595  0.0762707839 0.1874646 0.3001965
#> [11,] 0.8041906 0.9020410 0.8614809 0.7142903  0.0674209678 0.6753921 0.6399170
#> [12,] 0.4533880 0.5463373 0.7764817 0.8092277  0.0700493773 0.7601834 0.7817554
#> [13,] 0.4309508 0.4070249 0.3551160 0.3223111  0.0444856674 0.4045467 0.3264327
#> [14,] 0.9432231 0.9576202 0.8799696 0.6807167  0.0363335315 0.4927000 0.3699793
#> [15,] 0.6317660 0.7167340 0.8323695 0.8736814  0.0208609481 0.8337286 0.7653833
#> [16,] 0.1703104 0.4079189 0.6395920 0.7960714 -0.0136556717 0.7361555 0.5230327
#> [17,] 0.7699810 0.7642627 0.8512577 0.9728356  0.0184509288 0.9346235 0.8023144
#> [18,] 0.6399967 0.6237793 0.6625232 0.4989114  0.0464745195 0.5678526 0.4663073
#> [19,] 0.8970293 0.9276509 0.9079154 0.9239973  0.0259090814 0.8745276 0.7631781
#> [20,] 0.7546914 0.6876326 0.7232278 0.7695553  0.0188841375 0.7384884 0.5577571
#> [21,] 0.7613679 0.7558984 0.5796851 0.4503211  0.0466679229 0.5153956 0.4985197
#> [22,] 0.9675204 0.9765735 0.9049862 0.8268217  0.0074476119 0.7276721 0.6794542
#> [23,] 0.2706016 0.4546646 0.6265193 0.6557245  0.0415539478 0.6783744 0.7637256
#> [24,] 0.9111662 0.8911029 0.8160190 0.7822875  0.0182862570 0.6572306 0.5556011
#> [25,] 0.8666114 0.8839780 0.5060236 0.2813681  0.0396075968 0.1809520 0.3241005
#>             V32       V33       V34        V35        V36        V37       V38
#>  [1,] 0.3304826 0.5236703 0.6511356 0.81912583 0.82817645 0.77366554 0.8738996
#>  [2,] 0.3834632 0.7302201 0.8696190 0.95731134 0.92640065 0.90581734 0.9357793
#>  [3,] 0.4991371 0.5263786 0.6727980 0.83553020 0.80790863 0.80580599 0.8873832
#>  [4,] 0.5224342 0.3597580 0.1431917 0.02594654 0.10581051 0.30153710 0.4249883
#>  [5,] 0.8474345 0.7398098 0.6672224 0.66475463 0.70349823 0.67480160 0.5420313
#>  [6,] 0.2160411 0.3101885 0.3818169 0.40989761 0.37987222 0.37609940 0.4474917
#>  [7,] 0.8606743 1.0240305 0.9639663 0.68461082 0.37542805 0.05333765 0.2743827
#>  [8,] 0.5226795 0.6184947 0.4531997 0.22192479 0.14785502 0.23422352 0.1719097
#>  [9,] 0.5943420 0.4042302 0.2005754 0.11766676 0.01208238 0.12591201 0.1180980
#> [10,] 0.3091651 0.3215147 0.2800331 0.35438988 0.04740523 0.15532717 0.2030470
#> [11,] 0.6992944 0.5429739 0.3353302 0.15153462 0.22954557 0.28680936 0.4025026
#> [12,] 0.7013465 0.4987256 0.5078185 0.54382315 0.33370660 0.15987306 0.3201122
#> [13,] 0.4057927 0.4533277 0.3631890 0.23833624 0.25585540 0.28638871 0.3175928
#> [14,] 0.2992215 0.2562904 0.1913380 0.14825312 0.18599049 0.20094709 0.2469741
#> [15,] 0.6293941 0.5198354 0.4651529 0.52275099 0.57158859 0.48695020 0.3333328
#> [16,] 0.1667362 0.4700633 0.4627587 0.22092069 0.33834354 0.47166535 0.5813890
#> [17,] 0.7610496 0.8093772 0.6793644 0.55788952 0.46779488 0.35846084 0.3513598
#> [18,] 0.2237185 0.2577993 0.1719356 0.50641567 0.73905190 0.77321542 0.4252063
#> [19,] 0.6871765 0.7395005 0.8525770 0.97272530 0.99385812 0.93560764 0.8175420
#> [20,] 0.3804127 0.6256973 0.9277615 0.86083847 0.65578575 0.33415607 0.2895992
#> [21,] 0.3092266 0.2438249 0.2089884 0.49616259 0.60075098 0.53410035 0.2692692
#> [22,] 0.6038560 0.5499466 0.5652197 0.54569011 0.47421563 0.43739017 0.3479993
#> [23,] 0.6206395 0.6710390 0.8707535 0.97843416 0.99281944 0.78862603 0.6301438
#> [24,] 0.4632579 0.4345383 0.4921996 0.49260656 0.44712829 0.37525605 0.2776674
#> [25,] 0.5330852 0.5058890 0.3965055 0.39209850 0.40317461 0.52947024 0.4355863
#>             V39           V4        V40        V41        V42        V43
#>  [1,] 0.8365554  0.041457753 0.58988415 0.27991371 0.10506122 0.13675531
#>  [2,] 0.7130429  0.095029984 0.46811053 0.30392140 0.25710454 0.16343290
#>  [3,] 0.7683281  0.134703995 0.51876855 0.39442422 0.45379915 0.45573420
#>  [4,] 0.3613304  0.015202430 0.36963964 0.29857569 0.28169964 0.23364196
#>  [5,] 0.4564434  0.050532091 0.48710308 0.48120269 0.43303025 0.45389206
#>  [6,] 0.4108872  0.077706999 0.33732652 0.30054771 0.32272753 0.32322481
#>  [7,] 0.5871624  0.027450268 0.80118043 0.77546370 0.56512497 0.26947721
#>  [8,] 0.1921186  0.045611857 0.29014139 0.32246630 0.47797573 0.38993667
#>  [9,] 0.1402789  0.058838260 0.09736021 0.17010503 0.15732675 0.13447341
#> [10,] 0.1854390  0.122733690 0.12826596 0.06224847 0.13345864 0.12647239
#> [11,] 0.5368837  0.040034912 0.44301936 0.48223224 0.60249749 0.55124212
#> [12,] 0.2867595  0.066684270 0.15217847 0.52624696 0.68024448 0.57113946
#> [13,] 0.3038474  0.063574406 0.28047254 0.22711393 0.24825250 0.21429615
#> [14,] 0.2575789  0.051063522 0.22331961 0.17320161 0.21653232 0.25601628
#> [15,] 0.2711259  0.024756368 0.22555996 0.24263398 0.18388726 0.13891448
#> [16,] 0.6711090 -0.002873718 0.53122726 0.21311641 0.20887074 0.16662255
#> [17,] 0.3498064  0.028751488 0.33239514 0.22855891 0.14087290 0.15361856
#> [18,] 0.1220041  0.027902227 0.33346849 0.44357506 0.41587622 0.29470694
#> [19,] 0.6458320  0.032773327 0.46546519 0.29386831 0.18884699 0.15175809
#> [20,] 0.2024192  0.043783674 0.33464333 0.36208864 0.38128394 0.39633586
#> [21,] 0.2153878  0.023729563 0.39160174 0.35082102 0.30547230 0.22276907
#> [22,] 0.1683223  0.036977339 0.14847860 0.14448494 0.04013071 0.02724838
#> [23,] 0.5555489  0.014474284 0.55231130 0.47674742 0.43984845 0.31747809
#> [24,] 0.1936198  0.019195078 0.21179705 0.18103533 0.10251685 0.07270190
#> [25,] 0.5390070  0.005676542 0.52783755 0.34140564 0.28913115 0.25547236
#>               V44          V45           V46        V47        V48         V49
#>  [1,]  0.25892184  0.498927406  0.3968348720 0.21112005 0.17969664 0.108407915
#>  [2,]  0.24314927  0.384402916  0.2697354931 0.16154571 0.11739822 0.058010211
#>  [3,]  0.42869443  0.537108010  0.4181345968 0.08980327 0.09725819 0.040747206
#>  [4,]  0.27231508  0.216895722  0.1293997653 0.11082628 0.08126866 0.041592064
#>  [5,]  0.51892948  0.480118031  0.3454674134 0.24892244 0.19125072 0.105056014
#>  [6,]  0.24018820  0.131879569  0.0678729864 0.06316186 0.02971144 0.031911270
#>  [7,]  0.14393329  0.251496372  0.2575616157 0.20238442 0.16113731 0.125434363
#>  [8,]  0.15774685  0.184830871  0.1202126187 0.15893137 0.15014721 0.075375784
#>  [9,]  0.12108247  0.094388877  0.1242723054 0.14780814 0.11608187 0.056668613
#> [10,]  0.07865229  0.091954009  0.0570946983 0.04852457 0.06004498 0.034001784
#> [11,]  0.48307567  0.569993275  0.4077964951 0.26984966 0.22230988 0.129146832
#> [12,]  0.47403722  0.550856679  0.6342997113 0.46940287 0.27317605 0.142224160
#> [13,]  0.15028215  0.128440651  0.0963649882 0.12002782 0.11070308 0.063681410
#> [14,]  0.26417761  0.217491128  0.1304996059 0.09991994 0.08080576 0.048571279
#> [15,]  0.11523542  0.118567997  0.0772030571 0.05508711 0.05172078 0.028816282
#> [16,]  0.05711222 -0.001988249 -0.0185469233 0.06582455 0.07548197 0.018442885
#> [17,]  0.16849368  0.166334706  0.0930859202 0.05355693 0.05053046 0.023822961
#> [18,]  0.21910092  0.254719840  0.2620593078 0.23105111 0.19627292 0.134210696
#> [19,]  0.10404900  0.175082743  0.1932047764 0.11038742 0.06094561 0.017493710
#> [20,]  0.20998983  0.071739864  0.0530252094 0.09718400 0.06764231 0.026840784
#> [21,]  0.16278368  0.175322946  0.1507137699 0.13652402 0.11987136 0.071890439
#> [22,] -0.01017124 -0.007546721 -0.0003799833 0.01232237 0.01889524 0.007310021
#> [23,]  0.21210626  0.142923816  0.0755568759 0.09145698 0.08728487 0.053757671
#> [24,]  0.05423388  0.043656295  0.0492175470 0.02064032 0.01837909 0.006472201
#> [25,]  0.22507936  0.115899995  0.0839369804 0.08392476 0.04571981 0.047363060
#>                V5          V50          V51          V52         V53
#>  [1,] 0.031912939  0.050518952 0.0223021178 0.0093115937 0.006862096
#>  [2,] 0.064120097  0.020652350 0.0267345222 0.0196716370 0.013012156
#>  [3,] 0.129032901  0.023758588 0.0154063656 0.0413339736 0.025996730
#>  [4,] 0.039337045 -0.000102987 0.0140457777 0.0005627498 0.011932173
#>  [5,] 0.079957630  0.031078828 0.0162298024 0.0132628247 0.009466781
#>  [6,] 0.087701903  0.014958737 0.0126560613 0.0089872046 0.015749058
#>  [7,] 0.055567021  0.049075491 0.0040222380 0.0061789806 0.001672385
#>  [8,] 0.060704921  0.011922986 0.0215567401 0.0104453523 0.011404155
#>  [9,] 0.070639565  0.018242280 0.0165450903 0.0118265652 0.008021354
#> [10,] 0.063213593  0.021091822 0.0309802681 0.0241133326 0.025083882
#> [11,] 0.104939916  0.044677268 0.0282389641 0.0176201585 0.010724759
#> [12,] 0.055931308  0.050033234 0.0775136382 0.0552229303 0.023765193
#> [13,] 0.056167804  0.020158857 0.0143392331 0.0113318902 0.004984045
#> [14,] 0.099847914  0.016186793 0.0177888970 0.0174890363 0.013191688
#> [15,] 0.028813429  0.013295874 0.0038019298 0.0064309768 0.003095913
#> [16,] 0.031422564  0.011761066 0.0060850093 0.0220828841 0.028458225
#> [17,] 0.026580925  0.009646464 0.0077641049 0.0082848991 0.011344475
#> [18,] 0.031349718  0.036808009 0.0130687968 0.0153742827 0.011592777
#> [19,] 0.026588530  0.024283922 0.0254579554 0.0186065256 0.008058359
#> [20,] 0.065202449  0.021455944 0.0330647747 0.0165486384 0.013902294
#> [21,] 0.033752930  0.019284137 0.0186079472 0.0099163897 0.004360527
#> [22,] 0.052043558  0.009336392 0.0066670776 0.0051851611 0.006541897
#> [23,] 0.021280687  0.014948073 0.0069128453 0.0063794318 0.017495210
#> [24,] 0.043604974  0.005192673 0.0006292723 0.0044570396 0.005732440
#> [25,] 0.003071269  0.022535302 0.0076264848 0.0077119100 0.003614160
#>               V54         V55         V56           V57         V58         V59
#>  [1,] 0.013433560 0.017143144 0.013308363  1.755162e-02 0.015457861 0.010013604
#>  [2,] 0.027052836 0.023947164 0.012301135  1.202188e-02 0.013623987 0.013245867
#>  [3,] 0.013825462 0.013753295 0.018875886  1.130834e-02 0.016997354 0.013740114
#>  [4,] 0.011587733 0.010658010 0.009908484  4.128502e-03 0.001569018 0.008743609
#>  [5,] 0.007892572 0.005515142 0.008065445  6.609283e-03 0.004288504 0.002464491
#>  [6,] 0.020797049 0.014715646 0.016539710  1.371359e-02 0.015378825 0.018770587
#>  [7,] 0.002317821 0.008769941 0.002174921  7.700012e-03 0.010505283 0.008961582
#>  [8,] 0.009640062 0.001879923 0.004297711  2.846805e-03 0.008249803 0.006448251
#>  [9,] 0.008766003 0.005043446 0.005903621  5.563461e-03 0.006020269 0.006232112
#> [10,] 0.010596577 0.016612166 0.017727222  5.264452e-03 0.033560967 0.013548012
#> [11,] 0.011107650 0.015960126 0.010092747  9.120229e-03 0.009246830 0.007934468
#> [12,] 0.027169153 0.020815031 0.010715001  1.592331e-02 0.017862689 0.018587624
#> [13,] 0.011547130 0.005036717 0.006603999  4.293363e-03 0.009552780 0.003125265
#> [14,] 0.013151997 0.004973304 0.008148713  6.595384e-03 0.004975717 0.005944739
#> [15,] 0.005877716 0.008259339 0.004711392  8.940951e-03 0.007181342 0.005738384
#> [16,] 0.015995737 0.025380496 0.002940392 -4.730001e-05 0.004169845 0.008487617
#> [17,] 0.005466847 0.004930797 0.007284829  3.435046e-03 0.001235327 0.006774025
#> [18,] 0.008495259 0.003384926 0.005267301  2.259312e-02 0.009261164 0.004154499
#> [19,] 0.007662025 0.009025160 0.007669425  7.744030e-03 0.010175715 0.006212580
#> [20,] 0.012350022 0.004165543 0.009761748  1.166888e-02 0.003773149 0.006287796
#> [21,] 0.010980777 0.009335970 0.004252758  1.010070e-02 0.006583637 0.011218571
#> [22,] 0.003708691 0.003456088 0.004462596  4.123455e-03 0.003281359 0.002880006
#> [23,] 0.014155008 0.009980556 0.005628097  9.379532e-03 0.014796257 0.019591815
#> [24,] 0.006393076 0.007582705 0.006741078  5.438195e-03 0.006623636 0.005698357
#> [25,] 0.016588579 0.015006032 0.006731749  9.810010e-03 0.013656153 0.007709056
#>               V6          V60         V7         V8         V9
#>  [1,] 0.12534214 0.0075816045 0.20836452 0.30144964 0.46623815
#>  [2,] 0.11565754 0.0146405186 0.30659919 0.45825705 0.58544457
#>  [3,] 0.13316452 0.0164179728 0.16356807 0.38416824 0.65191303
#>  [4,] 0.07773671 0.0033335072 0.10419648 0.13292859 0.16838782
#>  [5,] 0.10484666 0.0031269825 0.14146576 0.14627220 0.27664012
#>  [6,] 0.09732093 0.0102096176 0.14102821 0.13610451 0.10117046
#>  [7,] 0.10934072 0.0034798536 0.11377293 0.14153492 0.21956317
#>  [8,] 0.07290369 0.0005440008 0.10802436 0.10518105 0.17842763
#>  [9,] 0.13102468 0.0081333612 0.12996033 0.17165353 0.21425581
#> [10,] 0.11710757 0.0159127062 0.17636379 0.12786900 0.27159924
#> [11,] 0.12057308 0.0043921192 0.11605728 0.14592810 0.24800225
#> [12,] 0.10740908 0.0091938655 0.09353629 0.15657295 0.17363569
#> [13,] 0.04660111 0.0047524462 0.10182814 0.12686093 0.24146244
#> [14,] 0.14462507 0.0046730149 0.16302805 0.19749064 0.26089207
#> [15,] 0.07394421 0.0036984574 0.08106313 0.10823655 0.10823001
#> [16,] 0.04403523 0.0085985917 0.02721170 0.06680396 0.14083084
#> [17,] 0.02900481 0.0044516974 0.04298422 0.03430684 0.05185573
#> [18,] 0.05094398 0.0032018281 0.19989415 0.29173821 0.26534211
#> [19,] 0.10159664 0.0103622604 0.14179004 0.07978601 0.07304339
#> [20,] 0.07952888 0.0092787520 0.10973064 0.10065975 0.08437163
#> [21,] 0.11466559 0.0082191721 0.11721258 0.20151095 0.27067855
#> [22,] 0.08177535 0.0021209654 0.08759449 0.09809261 0.16774509
#> [23,] 0.04561035 0.0008089812 0.04511475 0.05286955 0.01670733
#> [24,] 0.07728613 0.0045595751 0.09030300 0.09031622 0.10796658
#> [25,] 0.01987637 0.0001555185 0.02439896 0.02492619 0.04510402
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623