Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11       V12        V13       V14
#>  [1,] 0.037439724  0.14642159 0.25254693 0.3455713 0.41659763 0.4684383
#>  [2,] 0.065593385  0.12447855 0.24375304 0.2303241 0.28598494 0.3062896
#>  [3,] 0.032452481  0.29459904 0.27209931 0.2866817 0.30311280 0.2743609
#>  [4,] 0.019238069  0.25073323 0.29519600 0.3202234 0.38917404 0.5039805
#>  [5,] 0.030801997  0.24634647 0.30128025 0.2954579 0.26609534 0.1746684
#>  [6,] 0.091246898  0.33982539 0.49491076 0.5471963 0.61343892 0.6908450
#>  [7,] 0.034546078 -0.06591856 0.15163477 0.2744845 0.36202901 0.4178344
#>  [8,] 0.087608637  0.57955587 0.46993112 0.2035713 0.21688147 0.3186385
#>  [9,] 0.023678251  0.24126022 0.28719774 0.3099372 0.29584778 0.2438092
#> [10,] 0.043528254  0.18270296 0.21990379 0.3715879 0.49330879 0.5585595
#> [11,] 0.029003908  0.23453640 0.28549110 0.2735811 0.16191618 0.0326919
#> [12,] 0.032661087  0.39629878 0.46923625 0.5056664 0.52063193 0.4320613
#> [13,] 0.012900000  0.22450000 0.24310000 0.3134000 0.32060000 0.2917000
#> [14,] 0.044385313  0.05921888 0.13456204 0.1887768 0.23004035 0.2981838
#> [15,] 0.011342340  0.10169299 0.09921859 0.1312810 0.14188484 0.1263053
#> [16,] 0.016465377  0.17227427 0.23391453 0.1888466 0.15383225 0.2074052
#> [17,] 0.009161158  0.05707482 0.08880189 0.1079533 0.10464397 0.1711902
#> [18,] 0.013052947  0.15626088 0.14584427 0.1019872 0.09109199 0.1223272
#> [19,] 0.023408264  0.40025171 0.31211793 0.1310897 0.28884156 0.5389612
#> [20,] 0.032581021  0.31163597 0.23330268 0.2047011 0.23908147 0.1474932
#> [21,] 0.019829526  0.10145099 0.14329052 0.1666629 0.23993109 0.3500640
#> [22,] 0.013379117  0.09635249 0.09265431 0.1686059 0.18865671 0.1294815
#> [23,] 0.019283653  0.09937698 0.07236106 0.1158396 0.23357333 0.2520838
#> [24,] 0.020238604  0.10470190 0.11291606 0.1114293 0.11672488 0.1556611
#> [25,] 0.007232332  0.13888513 0.16093012 0.2074280 0.29643996 0.3170521
#>              V15        V16       V17       V18       V19           V2
#>  [1,] 0.60107800 0.65528222 0.7810164 0.8543261 0.9004706  0.049648176
#>  [2,] 0.37463931 0.44481734 0.4386876 0.2917198 0.3580099  0.116089200
#>  [3,] 0.25110881 0.26015617 0.2524463 0.2090590 0.2776992  0.034248747
#>  [4,] 0.53638512 0.56353852 0.6388942 0.7189710 0.8013694  0.033850154
#>  [5,] 0.15978095 0.17309281 0.2104859 0.2600060 0.2455454  0.037873147
#>  [6,] 0.82667343 0.93352222 0.9006286 0.9043711 0.9195685  0.107956757
#>  [7,] 0.39915856 0.50187090 0.7779755 0.7747035 0.7774854  0.035912712
#>  [8,] 0.53941996 0.55028600 0.1877172 0.1458996 0.3823928  0.095261655
#>  [9,] 0.23770541 0.30454351 0.3582592 0.4870192 0.6354470  0.044787937
#> [10,] 0.61623683 0.64796739 0.7279072 0.8022096 0.9028804  0.048584847
#> [11,] 0.08206135 0.14964368 0.2090796 0.2751079 0.3005392  0.035351163
#> [12,] 0.29587170 0.23905917 0.2741342 0.2937127 0.4560720  0.034285733
#> [13,] 0.22490000 0.23470000 0.2143000 0.2939000 0.4898000  0.014100000
#> [14,] 0.38001581 0.43920839 0.5732657 0.6934453 0.8527819  0.058009492
#> [15,] 0.11033419 0.22140269 0.3255034 0.3469058 0.3447006  0.014551659
#> [16,] 0.24489039 0.25013385 0.2233930 0.1059734 0.4259999  0.002477490
#> [17,] 0.11271971 0.09849075 0.1677821 0.2842540 0.2080613  0.006024664
#> [18,] 0.12083265 0.13448914 0.2243643 0.2816739 0.3630756  0.002951485
#> [19,] 0.66666228 0.85518767 0.9016863 0.8256632 0.5760677  0.058737441
#> [20,] 0.38144804 0.34164935 0.2279854 0.2141845 0.2557762 -0.002680979
#> [21,] 0.50470484 0.64932207 0.7601827 0.8023068 0.8235765  0.020133352
#> [22,] 0.15219529 0.14697764 0.1591508 0.2402258 0.1794326  0.033924357
#> [23,] 0.22519691 0.20087370 0.2285388 0.3260981 0.4400230  0.020177615
#> [24,] 0.18786542 0.19245119 0.1647225 0.1715527 0.1383504  0.019911816
#> [25,] 0.35122611 0.45043482 0.4720407 0.5133510 0.4775850  0.004360213
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.8745521 0.8650421 0.8511702 0.8136541 0.6571858 0.3999575 0.2510712
#>  [2,] 0.5166530 0.6617327 0.7247091 0.6366304 0.6534347 0.6301989 0.6515193
#>  [3,] 0.2964575 0.2882530 0.2941006 0.4093719 0.4777938 0.5878396 0.6945516
#>  [4,] 0.8463495 0.8985615 0.9438993 0.9217186 0.9195336 0.8891223 0.8805349
#>  [5,] 0.2310161 0.3649674 0.5728459 0.7211629 0.8004221 0.8629923 0.8908817
#>  [6,] 0.9468048 0.8909763 0.7002926 0.6012652 0.5741306 0.4865091 0.3914721
#>  [7,] 0.8122410 0.7971998 0.8264116 0.9676912 1.0634496 0.9580485 0.6974204
#>  [8,] 0.8834778 0.8920719 0.3828851 0.1524231 0.1493874 0.2051611 0.3574965
#>  [9,] 0.7139728 0.7953304 0.8005781 0.8704904 0.9623566 0.9536242 0.9551139
#> [10,] 0.8766500 0.9674783 0.8774945 0.8106373 0.8823391 0.7180138 0.6196590
#> [11,] 0.4328589 0.5655655 0.7019777 0.7372985 0.7853166 0.9249662 0.9962368
#> [12,] 0.6744120 0.8127767 0.8057347 0.7001899 0.6743418 0.6556417 0.8052787
#> [13,] 0.6127000 0.7531000 0.7718000 0.7432000 0.8673000 0.9308000 0.9836000
#> [14,] 0.9169908 0.9037404 0.9666748 0.9661285 0.9197906 0.8473358 0.7340513
#> [15,] 0.3725631 0.4658631 0.6296827 0.7217097 0.7563695 0.7810629 0.8290754
#> [16,] 0.6691339 0.5214946 0.5140272 0.9901398 0.8897807 0.5118043 0.5950073
#> [17,] 0.3344086 0.4567502 0.5898837 0.7297180 0.7913298 0.8752023 0.8828445
#> [18,] 0.5400913 0.6622314 0.8057469 0.9336121 0.8955481 0.9008013 0.9796472
#> [19,] 0.4310828 0.5076733 0.5938853 0.7414215 0.8672498 0.8119769 0.7120097
#> [20,] 0.4314959 0.3143858 0.6581395 0.5862438 0.2890917 0.1873925 0.5393316
#> [21,] 0.8178072 0.8795064 0.8982329 0.8442639 0.8399787 0.8537304 0.8613718
#> [22,] 0.2081159 0.2331781 0.1718620 0.1175546 0.1240535 0.1468286 0.3131617
#> [23,] 0.6038037 0.6112059 0.6514484 0.6495093 0.6530980 0.8026103 0.9269235
#> [24,] 0.1589557 0.2458790 0.2766043 0.4145957 0.6389024 0.8868744 0.9708856
#> [25,] 0.5134157 0.6459540 0.4936813 0.3774084 0.6267772 0.8313865 0.8344680
#>             V27       V28       V29           V3       V30       V31       V32
#>  [1,] 0.2180951 0.1983054 0.2458247  0.058310214 0.2929741 0.2617355 0.3498173
#>  [2,] 0.7388706 0.8082637 0.7954521  0.124120931 0.8563084 0.8986490 0.7882596
#>  [3,] 0.7356089 0.8344196 0.8505040  0.039539713 0.8574384 0.8566630 0.7261013
#>  [4,] 0.8390117 0.7112599 0.6105711  0.039869545 0.4825091 0.3369408 0.2155976
#>  [5,] 0.9243297 0.9784973 0.8629418  0.033090178 0.6680122 0.6723284 0.6064300
#>  [6,] 0.3892151 0.4100540 0.3581586  0.101694322 0.2663032 0.1922271 0.1617128
#>  [7,] 0.3520427 0.2056493 0.3488014  0.078292948 0.3967736 0.3896481 0.4257781
#>  [8,] 0.2602348 0.5718586 0.5519005  0.053829175 0.7233794 0.3709900 0.4053114
#>  [9,] 0.9559982 0.9655799 0.8534351  0.030791902 0.6389610 0.4371155 0.3143339
#> [10,] 0.5297909 0.3002300 0.2937770  0.045011504 0.3164684 0.1675368 0.0193107
#> [11,] 0.9840761 0.8187077 0.7082690  0.041408159 0.7374871 0.7025544 0.4873038
#> [12,] 0.9007482 0.8525527 0.7181876  0.053183698 0.5302921 0.4261982 0.4144137
#> [13,] 1.0000000 0.9595000 0.8722000  0.030900000 0.6862000 0.4901000 0.3280000
#> [14,] 0.6753382 0.5439458 0.3768285  0.131312115 0.3112777 0.2148309 0.1458205
#> [15,] 0.8415924 0.8677106 0.8992016  0.018802140 0.8634303 0.7632372 0.7129826
#> [16,] 0.6808858 0.7129938 0.5185393  0.038285334 0.3187996 0.2150065 0.1706353
#> [17,] 0.7843152 0.7165585 0.6972570  0.023617485 0.6178467 0.5687157 0.4909355
#> [18,] 0.9727998 0.9049151 0.8137436  0.009320292 0.6953629 0.5925930 0.5321012
#> [19,] 0.7530221 0.6129794 0.3358555  0.075575783 0.3876365 0.5676281 0.4729645
#> [20,] 0.8096444 0.6686179 0.7185624  0.008933613 0.9055215 0.5750925 0.3219821
#> [21,] 0.7626079 0.5811478 0.4873067  0.026201229 0.4788284 0.4179273 0.3051290
#> [22,] 0.3542477 0.4731146 0.5713522  0.026663846 0.6246597 0.6908796 0.7379395
#> [23,] 1.0174129 0.8728907 0.6539306  0.014108219 0.4838546 0.4830890 0.4169622
#> [24,] 0.8126605 0.6418433 0.5415730  0.016970147 0.3983394 0.2947899 0.1525985
#> [25,] 0.7498525 0.8301392 0.7374881 -0.003525469 0.5822870 0.5096527 0.4123280
#>             V33       V34        V35        V36        V37       V38        V39
#>  [1,] 0.3928466 0.4352206 0.40652939 0.35641800 0.31568099 0.3955707 0.33251723
#>  [2,] 0.4766685 0.4867029 0.58734186 0.30914357 0.33097070 0.3555797 0.30031585
#>  [3,] 0.5971861 0.4968179 0.41560199 0.39355427 0.37025101 0.3686532 0.37786916
#>  [4,] 0.1460913 0.1149992 0.14731644 0.18746686 0.20202865 0.2649723 0.30601715
#>  [5,] 0.4610345 0.2224743 0.12669339 0.13042046 0.16433465 0.1499237 0.18760173
#>  [6,] 0.1176071 0.1473372 0.22712085 0.10470872 0.12193136 0.1039810 0.14015009
#>  [7,] 0.5944546 0.5802868 0.46742005 0.25512208 0.23981794 0.4178132 0.42700485
#>  [8,] 0.6592714 0.8016657 0.89738791 0.86677131 0.85667696 0.9002827 0.72652910
#>  [9,] 0.2367902 0.1924916 0.18624659 0.11914982 0.08998398 0.1207246 0.16028569
#> [10,] 0.3169238 0.4346874 0.38536150 0.37458079 0.30276392 0.4378393 0.43817630
#> [11,] 0.2415506 0.1098956 0.09624611 0.06744323 0.17933556 0.1793174 0.12889374
#> [12,] 0.3808448 0.2788011 0.09250819 0.21399584 0.25575704 0.2746975 0.46491288
#> [13,] 0.3115000 0.1969000 0.10190000 0.03170000 0.07560000 0.0907000 0.10660000
#> [14,] 0.1714369 0.2599117 0.41930615 0.45144536 0.39316532 0.3959204 0.39445591
#> [15,] 0.7433691 0.7583676 0.75840076 0.66063395 0.54097044 0.4508009 0.33717402
#> [16,] 0.1356934 0.3623179 0.31092677 0.28066262 0.23904959 0.2262735 0.25795659
#> [17,] 0.5775205 0.6365735 0.61910852 0.53658679 0.37839161 0.2579482 0.12223653
#> [18,] 0.4972943 0.5283671 0.54941977 0.52273738 0.48227387 0.3715094 0.19064743
#> [19,] 0.2468977 0.3029644 0.67263009 0.68179082 0.55807933 0.4811177 0.36268398
#> [20,] 0.7210039 0.9288988 0.80497533 0.76266563 0.56750705 0.4559837 0.04674024
#> [21,] 0.2987761 0.3025421 0.19623707 0.10830072 0.14383293 0.1345777 0.15427818
#> [22,] 0.7388113 0.7760735 0.86419924 0.87406932 0.87627796 0.8436769 0.79372838
#> [23,] 0.2857822 0.1644035 0.17373588 0.27503342 0.27732859 0.3083439 0.30587429
#> [24,] 0.1581231 0.1893631 0.23081094 0.20166819 0.09446861 0.1084633 0.15185565
#> [25,] 0.2449045 0.1636878 0.22691067 0.34434883 0.22435980 0.2576716 0.27052612
#>                 V4       V40        V41        V42        V43        V44
#>  [1,]  0.087808259 0.2138390 0.16982172 0.21408616 0.23945970 0.14561238
#>  [2,]  0.131968352 0.1512135 0.56582717 0.69274254 0.43950474 0.33347239
#>  [3,]  0.047967070 0.4455468 0.52223519 0.50135079 0.49819400 0.47729052
#>  [4,]  0.060695996 0.2641413 0.24344525 0.23535867 0.16493499 0.16058778
#>  [5,]  0.050843452 0.1377089 0.18954240 0.16479105 0.15355344 0.13122817
#>  [6,]  0.069300629 0.1291453 0.09200279 0.13946834 0.19843390 0.19285415
#>  [7,]  0.038993014 0.4877493 0.52265383 0.49862398 0.22987002 0.16334130
#>  [8,]  0.104193345 0.4964287 0.32071572 0.29716332 0.24810946 0.29297313
#>  [9,]  0.022608997 0.1759872 0.11587159 0.11724468 0.17906180 0.19213773
#> [10,]  0.094940422 0.3066338 0.43909554 0.46708691 0.31595207 0.30804612
#> [11,]  0.053957632 0.0853586 0.05892123 0.15342687 0.17406919 0.15961210
#> [12,]  0.047741841 0.3998929 0.15286112 0.29083705 0.37240309 0.41633841
#> [13,]  0.037500000 0.1380000 0.06650000 0.14750000 0.24700000 0.27880000
#> [14,]  0.191730080 0.3402566 0.37511645 0.47766032 0.36223456 0.18561677
#> [15,]  0.027879382 0.2860582 0.23606048 0.16367734 0.15280288 0.13554892
#> [16,]  0.025514958 0.1312859 0.05525470 0.03782693 0.07746404 0.09815169
#> [17,]  0.023294122 0.2436416 0.26731738 0.27680459 0.26744372 0.16308263
#> [18,]  0.038695487 0.1626843 0.15647282 0.06278956 0.06630570 0.03857781
#> [19,]  0.062102883 0.3369006 0.22737561 0.22806562 0.25926020 0.23883206
#> [20,]  0.039297214 0.6298817 0.77152974 0.61586165 0.31957764 0.15530981
#> [21,]  0.041389171 0.1813071 0.15798929 0.10628235 0.08593182 0.10424133
#> [22,]  0.018903803 0.6796927 0.52523593 0.45886042 0.34010390 0.24496543
#> [23,]  0.004147658 0.2046344 0.12520449 0.22519828 0.19280141 0.14751925
#> [24,]  0.017683754 0.2153516 0.28160638 0.25168075 0.17597476 0.08994081
#> [25,] -0.010606342 0.1885940 0.09539728 0.11850416 0.11140748 0.09894466
#>              V45         V46        V47        V48          V49           V5
#>  [1,] 0.05196637  0.03153011 0.04066169 0.02661319  0.030351354  0.116903471
#>  [2,] 0.57927796  0.55741311 0.27283969 0.12263046  0.085384831  0.129496891
#>  [3,] 0.53801500  0.46009007 0.32338999 0.22534879  0.124258570  0.070238730
#>  [4,] 0.13814543  0.11227325 0.10227034 0.08490284  0.055322857  0.127431489
#>  [5,] 0.10860192  0.09785241 0.13383427 0.13320133  0.071035138  0.063364474
#>  [6,] 0.11458236  0.05641863 0.07444875 0.08180026  0.054937833  0.088694102
#>  [7,] 0.25353769  0.22719903 0.14433747 0.10520490  0.070571495  0.124707089
#>  [8,] 0.43069644  0.31618751 0.13727274 0.11060900  0.058176341  0.083520138
#>  [9,] 0.18699370  0.18128625 0.14516113 0.09442825  0.050958566  0.070762083
#> [10,] 0.26518006 -0.00108756 0.05854446 0.02750253 -0.005044816  0.017191494
#> [11,] 0.13317168  0.10944463 0.12138573 0.07753799  0.042206855  0.018987274
#> [12,] 0.39259637  0.17724465 0.10598701 0.12343907  0.073166971  0.091800557
#> [13,] 0.27090000  0.22830000 0.18180000 0.11850000  0.054600000  0.076700000
#> [14,] 0.16943281  0.13515688 0.09049778 0.07885916  0.045877032  0.248469467
#> [15,] 0.11454175  0.09542082 0.05203342 0.03887448  0.016983773  0.039780993
#> [16,] 0.08674548  0.06797602 0.04717324 0.01174410  0.018531712  0.048281084
#> [17,] 0.05482246  0.02698126 0.02613155 0.03927791  0.006894090  0.042495876
#> [18,] 0.03902972  0.03214009 0.02282873 0.01631353  0.010098944  0.051571238
#> [19,] 0.20748118  0.11100154 0.10131080 0.08685448  0.048956449  0.089122018
#> [20,] 0.20627555  0.33851091 0.23397491 0.05526635  0.005640826  0.071263736
#> [21,] 0.10469581  0.07535569 0.08388232 0.05798424  0.031059408  0.080718130
#> [22,] 0.24848381  0.20293469 0.16297102 0.11932487  0.064719700  0.003228463
#> [23,] 0.11570166  0.06422663 0.09775197 0.07143517  0.031036121  0.023872259
#> [24,] 0.09109197  0.09306824 0.04640492 0.03612998  0.031431221  0.048608868
#> [25,] 0.05101909  0.04952363 0.03433449 0.03396251  0.012558780 -0.003405693
#>               V50          V51          V52         V53          V54
#>  [1,] 0.012286477  0.012903881 0.0174531124 0.005856956 0.0191872902
#>  [2,] 0.038883936  0.032251032 0.0241305062 0.019684392 0.0153697780
#>  [3,] 0.038542948  0.023690828 0.0205048926 0.015601806 0.0105038377
#>  [4,] 0.022770050  0.019695611 0.0164601089 0.013809335 0.0207692700
#>  [5,] 0.019266852  0.017295522 0.0134568415 0.007032597 0.0110440699
#>  [6,] 0.017405024  0.033004724 0.0201788084 0.022551266 0.0187837418
#>  [7,] 0.003311645 -0.007286120 0.0053186529 0.009534388 0.0004643684
#>  [8,] 0.026381365  0.022230514 0.0233022058 0.017054115 0.0246263555
#>  [9,] 0.014384872  0.016891658 0.0122982916 0.010269101 0.0058677211
#> [10,] 0.003115172  0.023875948 0.0008989705 0.014832208 0.0113325529
#> [11,] 0.021241287  0.007246184 0.0097827840 0.006285674 0.0109197870
#> [12,] 0.017084347  0.023370680 0.0162140691 0.009150788 0.0077782021
#> [13,] 0.021900000  0.020400000 0.0124000000 0.009300000 0.0072000000
#> [14,] 0.016467487  0.023897026 0.0200095047 0.018696851 0.0212362689
#> [15,] 0.012946916  0.009117688 0.0102158687 0.007959667 0.0060721475
#> [16,] 0.014814810  0.027418828 0.0338724636 0.013674551 0.0010595346
#> [17,] 0.003881087  0.007444091 0.0030525894 0.010125637 0.0059280593
#> [18,] 0.010332970  0.006891608 0.0064213770 0.005451333 0.0034909007
#> [19,] 0.024191759  0.009983812 0.0171935709 0.006341449 0.0164394900
#> [20,] 0.047485359  0.046507044 0.0372403004 0.016713154 0.0166686176
#> [21,] 0.014053503  0.014286884 0.0061438845 0.007195940 0.0089355340
#> [22,] 0.028585173  0.022017920 0.0104804645 0.010822777 0.0060954894
#> [23,] 0.015299481  0.003038524 0.0028015444 0.004730416 0.0082560832
#> [24,] 0.010934745  0.010861690 0.0092867773 0.005821703 0.0075835970
#> [25,] 0.016124747  0.002891047 0.0069884900 0.004155134 0.0093321752
#>               V55         V56         V57           V58         V59         V6
#>  [1,] 0.016937000 0.012424901 0.009188717  0.0193950636 0.012032935 0.08800222
#>  [2,] 0.017099486 0.013796029 0.010869545  0.0126005379 0.016362094 0.12474914
#>  [3,] 0.007634059 0.005728752 0.007739292  0.0086338960 0.008178853 0.11526902
#>  [4,] 0.013151027 0.009127408 0.008339198  0.0104505829 0.007859580 0.17132940
#>  [5,] 0.005627147 0.008854427 0.008064509  0.0092786296 0.009094558 0.12872252
#>  [6,] 0.018835830 0.010231840 0.008723814  0.0143605897 0.012985056 0.16097234
#>  [7,] 0.007388377 0.010016718 0.005920457  0.0067852219 0.016766799 0.10291701
#>  [8,] 0.021388282 0.014456964 0.011703489  0.0141713849 0.012560956 0.11097303
#>  [9,] 0.006213934 0.006038998 0.004724252  0.0039254772 0.003889728 0.12350679
#> [10,] 0.005057398 0.021585965 0.016010372  0.0055229155 0.030064118 0.12197435
#> [11,] 0.003716189 0.004108611 0.003525955  0.0069388784 0.004099514 0.11459504
#> [12,] 0.006923777 0.004079352 0.003125499  0.0044958554 0.006538924 0.14749075
#> [13,] 0.001900000 0.002700000 0.005400000  0.0017000000 0.002400000 0.07870000
#> [14,] 0.006793437 0.012838500 0.014147218  0.0123352951 0.013119857 0.15753108
#> [15,] 0.005463902 0.005798060 0.008305358  0.0063921665 0.005025958 0.06281963
#> [16,] 0.009788824 0.010040508 0.007422679  0.0162270131 0.009263000 0.12948169
#> [17,] 0.003899708 0.007412743 0.007405462  0.0042342193 0.007100525 0.07485094
#> [18,] 0.003483995 0.004097540 0.003600755  0.0025507916 0.003170930 0.08036884
#> [19,] 0.011949400 0.004270113 0.008877723  0.0062509688 0.011081192 0.19373782
#> [20,] 0.009863898 0.003866027 0.009540547 -0.0002242608 0.006091071 0.11210589
#> [21,] 0.007890843 0.006895397 0.003590439  0.0033093416 0.003485861 0.11094949
#> [22,] 0.015111578 0.009398871 0.001076015  0.0063870013 0.002816461 0.05448307
#> [23,] 0.005110024 0.005933544 0.005159707  0.0078762082 0.005026006 0.04166643
#> [24,] 0.007392706 0.003436232 0.003348480  0.0025214614 0.003554455 0.06100327
#> [25,] 0.009058543 0.008172253 0.008389567  0.0042964381 0.003416823 0.08701505
#>                V60         V7          V8         V9
#>  [1,] 5.922701e-03 0.11587280 0.079981636 0.08913388
#>  [2,] 1.824887e-02 0.09239801 0.129005533 0.18170807
#>  [3,] 4.250359e-03 0.11412045 0.142402454 0.25385575
#>  [4,] 4.950791e-03 0.18614560 0.188420280 0.22768588
#>  [5,] 6.268170e-03 0.14140245 0.134982575 0.16608849
#>  [6,] 1.528389e-02 0.17059476 0.141860572 0.22687514
#>  [7,] 4.548067e-03 0.09409583 0.008336218 0.07891440
#>  [8,] 1.352902e-02 0.24183006 0.424107645 0.59157652
#>  [9,] 5.201227e-03 0.14274521 0.159893673 0.20080026
#> [10,] 2.427060e-02 0.21995641 0.141026656 0.04295661
#> [11,] 6.076520e-03 0.12887850 0.117651712 0.19197320
#> [12,] 5.069657e-03 0.12903323 0.162957385 0.27161752
#> [13,] 2.900000e-03 0.06620000 0.110800000 0.17770000
#> [14,] 7.669592e-03 0.07486929 0.170970951 0.07172597
#> [15,] 6.050665e-03 0.07231556 0.060998177 0.06271078
#> [16,] 5.622213e-03 0.12303142 0.192102801 0.15274597
#> [17,] 9.433475e-03 0.08039622 0.073405092 0.07881046
#> [18,] 2.367544e-03 0.09488983 0.089664611 0.16486997
#> [19,] 7.950030e-03 0.19957682 0.222063885 0.32620566
#> [20,] 5.355365e-03 0.17505507 0.145231199 0.13850249
#> [21,] 4.647994e-03 0.12055531 0.079347063 0.10353954
#> [22,] 1.662348e-05 0.07850578 0.110414491 0.02487424
#> [23,] 5.949815e-03 0.06718816 0.067112081 0.07298322
#> [24,] 2.408145e-03 0.07077978 0.068187518 0.10150118
#> [25,] 3.143377e-03 0.13289622 0.149794660 0.13332935
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986