Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1           V10        V11        V12        V13        V14
#>  [1,] 0.025610751  0.2420664002 0.28444890 0.33767312 0.34326826 0.38666856
#>  [2,] 0.029140604  0.1620633769 0.29684319 0.45434357 0.53385298 0.52293458
#>  [3,] 0.051215412  0.1625964716 0.21519138 0.28691869 0.33062554 0.28699053
#>  [4,] 0.022393826  0.3514494007 0.44097339 0.51031407 0.51377759 0.41825573
#>  [5,] 0.117255960  0.3211510623 0.31022627 0.18354080 0.24335429 0.29861424
#>  [6,] 0.084057075  0.1676554466 0.22677523 0.22572689 0.27052523 0.31154041
#>  [7,] 0.044037362  0.2342175157 0.29444878 0.25984680 0.38429770 0.56302288
#>  [8,] 0.049514923  0.4825231042 0.53840217 0.52257101 0.44616373 0.32924102
#>  [9,] 0.056250507 -0.0322653604 0.12812662 0.28156935 0.32488002 0.39463609
#> [10,] 0.025074271  0.3413625723 0.30985864 0.34105296 0.32476957 0.28434010
#> [11,] 0.019320771  0.0005682812 0.21877579 0.24997981 0.12610725 0.32567285
#> [12,] 0.024666030  0.2708125779 0.31450443 0.35750037 0.31155581 0.20513787
#> [13,] 0.024349763  0.2578528053 0.29910364 0.31662228 0.24492409 0.13773496
#> [14,] 0.016715412  0.2178649721 0.24491055 0.30866879 0.30113850 0.27001768
#> [15,] 0.009686537  0.1573740020 0.13875242 0.21542761 0.33635597 0.12721077
#> [16,] 0.020493674  0.1566294228 0.13126518 0.14352423 0.12888229 0.14261660
#> [17,] 0.012863901  0.1652215611 0.09918911 0.12845873 0.32287313 0.26942335
#> [18,] 0.014880011  0.0449647808 0.16301448 0.35394878 0.27928612 0.22546239
#> [19,] 0.007109330  0.0186953950 0.01971747 0.03578363 0.07398231 0.11836654
#> [20,] 0.011474807  0.1287738300 0.14113707 0.12708381 0.13899383 0.12544943
#> [21,] 0.025723273 -0.0178257857 0.02819072 0.25353717 0.13527479 0.09165864
#> [22,] 0.028212345  0.0902455375 0.14850454 0.06850623 0.13368954 0.14742193
#> [23,] 0.004845414  0.0171276732 0.02419832 0.07134642 0.11294891 0.10778264
#> [24,] 0.013182168  0.1091488947 0.16687275 0.16532536 0.16103120 0.17516326
#> [25,] 0.023315901  0.1683874075 0.18705478 0.17021161 0.24936036 0.37902600
#>              V15       V16       V17       V18       V19          V2
#>  [1,] 0.43717389 0.5199800 0.5999830 0.6874342 0.7822463 0.033682994
#>  [2,] 0.60324240 0.7114579 0.8952985 0.9764351 0.9771643 0.052315080
#>  [3,] 0.28999508 0.3884170 0.3809698 0.2040908 0.3386975 0.107374561
#>  [4,] 0.31083760 0.3050219 0.3279680 0.3163295 0.4452912 0.021146439
#>  [5,] 0.37266598 0.3897022 0.3558172 0.3682034 0.4639539 0.183936335
#>  [6,] 0.37820425 0.3395576 0.3902918 0.2302820 0.3181661 0.079197588
#>  [7,] 0.62961448 0.7481582 0.7554447 0.7962039 0.8868069 0.044515736
#>  [8,] 0.26439989 0.3114310 0.3320343 0.3647689 0.5014055 0.063200700
#>  [9,] 0.52547045 0.5358841 0.6675046 0.7843945 0.8944223 0.061431342
#> [10,] 0.24955515 0.2430729 0.2016083 0.1838610 0.2518264 0.025594398
#> [11,] 0.40260280 0.5244952 0.4529803 0.3758752 0.3376256 0.038782750
#> [12,] 0.16727236 0.2486848 0.2633671 0.3972343 0.5984267 0.024113991
#> [13,] 0.12944980 0.1685230 0.2136571 0.2474229 0.2832901 0.032048377
#> [14,] 0.20557079 0.2205535 0.2104892 0.2959691 0.4949308 0.019394632
#> [15,] 0.20840508 0.2496322 0.2898956 0.3196064 0.2527461 0.045800255
#> [16,] 0.14443943 0.2588485 0.4274353 0.4256298 0.3702171 0.020378143
#> [17,] 0.12739291 0.1690889 0.1092348 0.3449366 0.8189350 0.027028224
#> [18,] 0.09651792 0.0800372 0.1697388 0.4012038 0.2618413 0.044346975
#> [19,] 0.15860723 0.2642691 0.3331118 0.3595156 0.3526903 0.008515422
#> [20,] 0.19140636 0.2557788 0.3031748 0.3275430 0.2683301 0.017447013
#> [21,] 0.07967696 0.1162187 0.1203545 0.1417306 0.1345220 0.025079100
#> [22,] 0.15509891 0.1689418 0.1688379 0.1549439 0.1146752 0.019576846
#> [23,] 0.05742906 0.1003120 0.1137986 0.1513663 0.1544673 0.006641634
#> [24,] 0.26169111 0.2971110 0.2820949 0.2332139 0.2814555 0.022892444
#> [25,] 0.54234975 0.6991999 0.8164825 0.8094444 0.7891949 0.036448158
#>               V20        V21           V22       V23       V24       V25
#>  [1,]  0.84992778 0.85438579  0.9077920231 0.9354676 0.9551526 0.9307779
#>  [2,]  0.91061163 0.85205704  0.7322585710 0.5979173 0.5226788 0.3217086
#>  [3,]  0.43108127 0.56217466  0.5989984931 0.6071003 0.6593015 0.5594914
#>  [4,]  0.64393540 0.74886332  0.8280067732 0.7876602 0.7946906 0.8380596
#>  [5,]  0.50452349 0.43755298  0.2547263319 0.3330656 0.4961985 0.3891590
#>  [6,]  0.37029764 0.36997275  0.3399772121 0.3885413 0.2882016 0.4065551
#>  [7,]  0.92795648 0.98066493  0.9840842611 0.9712661 0.7878036 0.5335330
#>  [8,]  0.66754527 0.74565650  0.8514919717 0.8970851 0.9458780 0.9032273
#>  [9,]  0.98421529 0.94566772  0.9410148878 0.9374201 0.9780781 0.9253131
#> [10,]  0.32641288 0.35362109  0.3560106033 0.4684206 0.5700004 0.6556192
#> [11,]  0.24281823 0.21519439  0.0810737814 0.2530292 0.5230198 0.3617942
#> [12,]  0.71568322 0.80572165  0.7698282843 0.8485768 0.9616440 0.9227276
#> [13,]  0.32593099 0.44963746  0.5855793825 0.6914065 0.7534116 0.8052859
#> [14,]  0.61812293 0.75650656  0.7802465219 0.7534566 0.8730953 0.9243547
#> [15,]  0.04169816 0.09412878  0.1025939117 0.1761361 0.1790116 0.2555595
#> [16,]  0.44301337 0.55186543  0.6806542316 0.8043037 0.8115274 0.8002860
#> [17,]  0.99752155 0.85256625  0.7344871117 0.6063408 0.4990579 0.5392889
#> [18,]  0.55532712 0.72675272  0.7467043219 0.7628381 0.8195838 0.9890743
#> [19,]  0.32702085 0.38436872  0.4026474527 0.3627617 0.4410023 0.4655407
#> [20,]  0.23729227 0.33541307  0.3605089107 0.4522554 0.6587682 0.8399611
#> [21,] -0.00344423 0.10812424 -0.0009751438 0.2107447 0.2356239 0.0125763
#> [22,]  0.19421861 0.33991661  0.5683072803 0.8275608 0.8723405 0.8194176
#> [23,]  0.15807531 0.22427910  0.3551581183 0.3842219 0.5179019 0.5762149
#> [24,]  0.34140496 0.28977547  0.4385391058 0.5951265 0.6566684 0.7267177
#> [25,]  0.76209298 0.83809671  0.8795644728 0.8255238 0.8275611 0.8240685
#>              V26        V27        V28        V29          V3       V30
#>  [1,] 0.91549885 0.88750634 0.76268100 0.67434249 0.043850078 0.5568172
#>  [2,] 0.05250296 0.04331381 0.08847439 0.07505953 0.049135060 0.1305486
#>  [3,] 0.52820520 0.71435905 0.83465479 0.78037223 0.079344307 0.8249774
#>  [4,] 0.91268436 0.95793920 0.83776009 0.63351663 0.032690257 0.4818474
#>  [5,] 0.29162484 0.34686300 0.53025525 0.50505436 0.224214825 0.4810876
#>  [6,] 0.45735953 0.45387657 0.70815492 0.75688497 0.065047362 0.6834583
#>  [7,] 0.61873908 0.60766233 0.55300560 0.48771709 0.033761952 0.3804747
#>  [8,] 0.83359430 0.79904849 0.61976633 0.40660682 0.050133199 0.3280304
#>  [9,] 0.68275304 0.49124645 0.40512113 0.33394408 0.110249103 0.2559012
#> [10,] 0.78331404 0.81714677 0.86316944 0.83682938 0.039552623 0.8470941
#> [11,] 0.36971228 0.52203300 0.49109729 0.42745073 0.056856476 0.3869286
#> [12,] 0.92039961 0.95736210 0.98666949 0.86112733 0.036672864 0.6212878
#> [13,] 0.84970533 0.92451406 0.92619207 0.82866380 0.039122133 0.7156699
#> [14,] 0.97761857 1.00000000 0.96078440 0.86352813 0.034419769 0.6693551
#> [15,] 0.25283072 0.11430022 0.53900384 0.72674717 0.031288623 0.6336348
#> [16,] 0.84948007 0.90172590 0.85015956 0.81475465 0.024508762 0.7907288
#> [17,] 0.84321914 1.03242616 0.74664121 0.44946443 0.036001905 0.1271543
#> [18,] 0.92512078 0.74900861 0.57588295 0.53397461 0.059408733 0.4543356
#> [19,] 0.49663130 0.70707968 0.92000786 1.00941388 0.009182371 0.9488678
#> [20,] 0.85813073 0.75916219 0.70615687 0.57967260 0.015861978 0.3896685
#> [21,] 0.16475836 0.47889744 0.68113811 0.67552425 0.016632385 0.6109482
#> [22,] 0.83574752 0.80870517 0.85963008 0.94933613 0.027419745 0.8029321
#> [23,] 0.65483690 0.68631443 0.82074352 0.96848201 0.019514129 0.9001457
#> [24,] 0.73109893 0.67409384 0.74129987 0.72399848 0.025667259 0.5535468
#> [25,] 0.83092795 0.73849174 0.57397598 0.49174208 0.049536008 0.4927180
#>             V31       V32       V33       V34       V35        V36        V37
#>  [1,] 0.3702152 0.2537927 0.2014346 0.1975654 0.2249846 0.16240666 0.19313737
#>  [2,] 0.2327079 0.3688925 0.3753412 0.4803258 0.5144200 0.28162449 0.19173440
#>  [3,] 0.8102855 0.6952453 0.3965536 0.4654243 0.5704962 0.39326301 0.30480263
#>  [4,] 0.4311246 0.5133379 0.4599021 0.3081388 0.0781795 0.26956089 0.34904883
#>  [5,] 0.3265330 0.2737190 0.2907942 0.3761647 0.4990405 0.61773017 0.53986741
#>  [6,] 0.7281353 0.6654939 0.5651926 0.5971984 0.6110695 0.40406164 0.16552647
#>  [7,] 0.2771186 0.2442607 0.1833979 0.1646904 0.1960618 0.23916366 0.18733616
#>  [8,] 0.3651286 0.3284371 0.4181464 0.4503790 0.3022275 0.40107009 0.43587548
#>  [9,] 0.2610015 0.2971140 0.3691445 0.4907000 0.5513535 0.38418991 0.26280627
#> [10,] 0.8545688 0.7556039 0.5974426 0.4489352 0.3249630 0.35703430 0.38667256
#> [11,] 0.3457800 0.3809317 0.4999397 0.4147469 0.1968012 0.15141979 0.08889412
#> [12,] 0.3932479 0.2635881 0.1827855 0.1435686 0.1304559 0.04894969 0.06463664
#> [13,] 0.6516572 0.5178915 0.4051044 0.2908665 0.1724207 0.05852679 0.15867491
#> [14,] 0.4741376 0.3255132 0.3035583 0.1996340 0.1098413 0.03304608 0.07437773
#> [15,] 0.5210505 0.6428740 0.7284271 0.6203933 0.5674028 0.75183080 0.88099614
#> [16,] 0.7888334 0.7207173 0.6807796 0.6965079 0.6896158 0.64373065 0.61173060
#> [17,] 0.2442823 0.3909692 0.3581334 0.3712980 0.3955433 0.28352910 0.36291215
#> [18,] 0.4662773 0.4873364 0.6187333 0.5453136 0.5880787 0.63613070 0.58556293
#> [19,] 0.8724933 0.7630631 0.7056803 0.6058253 0.5749632 0.59339490 0.53751535
#> [20,] 0.2917338 0.2422883 0.2207260 0.2202987 0.2602091 0.22955841 0.16167071
#> [21,] 0.7568310 0.5901573 0.6728002 0.8880194 0.9765756 0.96705329 0.78133490
#> [22,] 0.6341561 0.5433832 0.5969236 0.6844406 0.6517298 0.48043179 0.25092125
#> [23,] 0.6902830 0.6437757 0.7532838 0.6030380 0.4617902 0.41888020 0.39923690
#> [24,] 0.4519606 0.3993878 0.5376941 0.8266555 0.8456070 0.72032160 0.44212789
#> [25,] 0.4353733 0.2808347 0.2738193 0.2777997 0.2463365 0.22549709 0.28285845
#>              V38        V39           V4        V40        V41          V42
#>  [1,] 0.25301815 0.27781265  0.074008334 0.27086448 0.30885501 0.3115450716
#>  [2,] 0.37777947 0.33776646  0.081729781 0.14020031 0.08090780 0.1514862321
#>  [3,] 0.35364802 0.31529927  0.065430491 0.23838872 0.60035743 0.6632785518
#>  [4,] 0.33303874 0.44537860  0.028215752 0.39417721 0.16745628 0.2562370604
#>  [5,] 0.53436788 0.52161660  0.316769561 0.46935050 0.37417809 0.2833958469
#>  [6,] 0.34067101 0.31281930  0.078538485 0.17779657 0.52135209 0.7417006946
#>  [7,] 0.20482341 0.28581436  0.050823788 0.23802494 0.14123504 0.2162637159
#>  [8,] 0.34388653 0.27499099  0.045270484 0.23739619 0.18763382 0.2904338497
#>  [9,] 0.33027405 0.45103207  0.099573055 0.43274171 0.31041802 0.3564418047
#> [10,] 0.38207080 0.44104650  0.032568223 0.51733486 0.54401488 0.5225904600
#> [11,] 0.20140790 0.32639326  0.085354716 0.37000417 0.29178461 0.3187563459
#> [12,] 0.08400597 0.09588613  0.044539259 0.10985020 0.07936548 0.1020269844
#> [13,] 0.15320390 0.14487435  0.054816585 0.17317341 0.23665115 0.2020085905
#> [14,] 0.10330783 0.10797832  0.036817497 0.13229473 0.05912510 0.1417086306
#> [15,] 0.93551750 0.97509946  0.025589865 0.93673523 0.91982250 0.8143302106
#> [16,] 0.53391312 0.36100474  0.040771282 0.25952941 0.19842711 0.0972130588
#> [17,] 0.39505041 0.31517640  0.013455489 0.34281658 0.23926606 0.1700184080
#> [18,] 0.29522418 0.04016432  0.049576858 0.32329805 0.40547430 0.5747397611
#> [19,] 0.35585413 0.14213575 -0.002121312 0.06260817 0.04679371 0.1444831077
#> [20,] 0.14122775 0.16968453  0.026522625 0.19879849 0.22604127 0.2016678927
#> [21,] 0.64343509 0.63195034  0.025085296 0.68917810 0.61279343 0.5319190653
#> [22,] 0.13323536 0.18278437  0.041649864 0.26530072 0.20162364 0.2005533277
#> [23,] 0.45507292 0.47545974  0.035713229 0.36140880 0.16656177 0.0002731032
#> [24,] 0.29827799 0.24898893  0.040159040 0.33990869 0.26562449 0.3398685164
#> [25,] 0.18566540 0.11637534  0.053441968 0.19487232 0.19528323 0.1660037057
#>              V43        V44        V45        V46         V47        V48
#>  [1,] 0.19090825 0.15578030 0.16091591 0.12509268 0.085588706 0.05618763
#>  [2,] 0.16153901 0.02380165 0.07472328 0.06313112 0.047827875 0.04645105
#>  [3,] 0.50119432 0.38276660 0.52855075 0.54947956 0.345966108 0.18645999
#>  [4,] 0.37552101 0.45548715 0.46261830 0.24880747 0.138138057 0.10791391
#>  [5,] 0.18982940 0.22639477 0.33613475 0.27644167 0.191134570 0.16298307
#>  [6,] 0.65525369 0.51559604 0.57593473 0.65885834 0.464966673 0.27272696
#>  [7,] 0.32501983 0.24707306 0.09661144 0.10497565 0.083384060 0.09743936
#>  [8,] 0.39981923 0.35799274 0.25276548 0.10728262 0.068170536 0.08868405
#>  [9,] 0.32840385 0.25389724 0.20161186 0.14755102 0.076754732 0.03155164
#> [10,] 0.52661942 0.52298894 0.60494055 0.47277618 0.339261430 0.24724939
#> [11,] 0.27080123 0.20088656 0.21966500 0.18599715 0.087342118 0.11293439
#> [12,] 0.18905542 0.21908131 0.19605556 0.21603638 0.180018378 0.11194525
#> [13,] 0.18151784 0.15127698 0.11786272 0.11420270 0.129593165 0.11896927
#> [14,] 0.23913097 0.26407183 0.25090233 0.22188785 0.178004296 0.11433833
#> [15,] 0.57826812 0.29293535 0.41118949 0.52394068 0.423907317 0.28138968
#> [16,] 0.08380036 0.06569779 0.05757524 0.07192785 0.043598208 0.03230909
#> [17,] 0.17414966 0.12490221 0.11419429 0.11426983 0.110492440 0.07946820
#> [18,] 0.51541862 0.33820372 0.19714673 0.12430570 0.068326994 0.06479168
#> [19,] 0.22687019 0.25860647 0.21519912 0.15197828 0.088202184 0.10872762
#> [20,] 0.16281112 0.13414301 0.11940705 0.11065466 0.051959592 0.04689593
#> [21,] 0.39359629 0.21683409 0.11296392 0.05159266 0.064109630 0.05556706
#> [22,] 0.27413320 0.17383100 0.05921008 0.06218231 0.049480316 0.04079586
#> [23,] 0.01703036 0.05887374 0.09329251 0.02471722 0.008945859 0.02986644
#> [24,] 0.30913028 0.25859127 0.11880848 0.09079975 0.094325619 0.06937402
#> [25,] 0.11812747 0.12435339 0.14314509 0.11911836 0.120633755 0.09016834
#>               V49            V5         V50          V51           V52
#>  [1,] 0.035005483  0.1262834685 0.021611140 0.0155932807  0.0164138449
#>  [2,] 0.038047904  0.1127354220 0.021317375 0.0181788636  0.0199872660
#>  [3,] 0.121830247  0.0721373235 0.057444451 0.0533619970  0.0344750913
#>  [4,] 0.052518855  0.0631888780 0.012643496 0.0165795924  0.0120263144
#>  [5,] 0.077371468  0.2900061848 0.025902411 0.0204751333  0.0292613806
#>  [6,] 0.133138761  0.0727557359 0.044234275 0.0792838045  0.0620431557
#>  [7,] 0.082968044  0.0950245796 0.022111254 0.0284358240  0.0198040125
#>  [8,] 0.050407987  0.0229533114 0.013168414 0.0154687691  0.0090037163
#>  [9,] 0.035425092  0.1754927496 0.006743221 0.0009957797  0.0110355640
#> [10,] 0.137006689  0.0737173662 0.041582684 0.0164599832  0.0122353393
#> [11,] 0.070176835  0.0531242194 0.015588234 0.0079782033  0.0197635693
#> [12,] 0.054916034  0.0656396877 0.014661762 0.0159005908  0.0136365832
#> [13,] 0.066691744  0.0589542748 0.014499271 0.0175545944  0.0137209303
#> [14,] 0.053499753  0.0755732466 0.020885525 0.0197797079  0.0120118082
#> [15,] 0.192052065  0.0639999777 0.075371980 0.0379962315  0.0169060645
#> [16,] 0.013944202  0.0614562715 0.014407716 0.0092643352  0.0094927023
#> [17,] 0.052114905  0.0519700379 0.031739752 0.0052223247  0.0127509397
#> [18,] 0.011821654  0.1097945045 0.004174391 0.0020514297 -0.0002884964
#> [19,] 0.065030945 -0.0009397212 0.015745178 0.0082231060  0.0077165887
#> [20,] 0.028803559  0.0411303368 0.011659919 0.0052422356  0.0114437998
#> [21,] 0.043493298  0.0745417401 0.016786484 0.0058935419  0.0063809933
#> [22,] 0.008040515  0.0544892542 0.007954963 0.0071578015  0.0055234808
#> [23,] 0.009516753  0.0268406110 0.003427382 0.0044024977  0.0056740309
#> [24,] 0.024026152  0.0927741036 0.019743664 0.0285809799  0.0103510703
#> [25,] 0.050552510  0.0764256513 0.019424668 0.0126330536  0.0082424797
#>               V53         V54         V55         V56         V57          V58
#>  [1,] 0.015879859 0.015991973 0.007884500 0.011187359 0.008281888 0.0065867592
#>  [2,] 0.012620140 0.021622599 0.022460992 0.007087919 0.002146626 0.0119952359
#>  [3,] 0.016434244 0.021406947 0.018610148 0.014665245 0.009060678 0.0166597385
#>  [4,] 0.008633872 0.009653409 0.008294713 0.005237582 0.004535339 0.0035289588
#>  [5,] 0.019304140 0.017836255 0.017955530 0.005804546 0.012389635 0.0129809136
#>  [6,] 0.031314161 0.027597992 0.020872033 0.009080895 0.017756930 0.0183845791
#>  [7,] 0.013307536 0.022697331 0.011740960 0.004683671 0.004372407 0.0163645701
#>  [8,] 0.007517618 0.012710707 0.008971565 0.008491679 0.006700856 0.0108480387
#>  [9,] 0.016430025 0.017669751 0.007600057 0.015937529 0.013730693 0.0162023243
#> [10,] 0.012556509 0.008342359 0.006271225 0.005872020 0.006039313 0.0060049893
#> [11,] 0.019795156 0.017174705 0.032269273 0.008773842 0.026743327 0.0247761887
#> [12,] 0.004810204 0.003866164 0.004716662 0.005142246 0.003880017 0.0024575951
#> [13,] 0.009468066 0.010482605 0.006511369 0.007290595 0.004974975 0.0053305766
#> [14,] 0.008373818 0.006966165 0.002553798 0.002589154 0.004726494 0.0016347231
#> [15,] 0.015081605 0.010091836 0.018847540 0.012323404 0.001633717 0.0162707381
#> [16,] 0.006963225 0.005851806 0.005418394 0.004566919 0.006783803 0.0072468491
#> [17,] 0.013828994 0.016280913 0.010455782 0.006379962 0.007097967 0.0106876044
#> [18,] 0.010674752 0.009445623 0.008714876 0.010179398 0.010606375 0.0055050364
#> [19,] 0.001526921 0.008931762 0.006984903 0.003587003 0.004403157 0.0035918820
#> [20,] 0.006870792 0.008517946 0.006703925 0.003893144 0.004031736 0.0046648136
#> [21,] 0.013603309 0.012420793 0.007690718 0.005470119 0.010585885 0.0043686584
#> [22,] 0.006641215 0.006809115 0.006138901 0.006949877 0.014175844 0.0111684246
#> [23,] 0.014415823 0.007790923 0.005689180 0.010176174 0.006076969 0.0005955507
#> [24,] 0.011049716 0.012189305 0.007389522 0.010507884 0.013306662 0.0037220590
#> [25,] 0.008131449 0.008955378 0.007243517 0.007776741 0.006608195 0.0039847229
#>                V59         V6         V60         V7          V8           V9
#>  [1,] 0.0062277246 0.14255457 0.005250167 0.15892173  0.16012895  0.207987002
#>  [2,] 0.0123292029 0.05433227 0.005817306 0.09836040  0.08455514  0.053564617
#>  [3,] 0.0145761177 0.13933107 0.011754167 0.13799740  0.19640458  0.230624785
#>  [4,] 0.0070693522 0.10682822 0.004221447 0.11031244  0.14036253  0.277537754
#>  [5,] 0.0168854194 0.14691381 0.021899155 0.19602674  0.10001025  0.281792178
#>  [6,] 0.0185559960 0.11121621 0.008558506 0.07987900  0.16465530  0.167266259
#>  [7,] 0.0103270123 0.14961089 0.004320762 0.16453910  0.15726865  0.200095268
#>  [8,] 0.0055002627 0.09315566 0.004588034 0.10176258  0.20510699  0.360872803
#>  [9,] 0.0129822554 0.18427845 0.005935345 0.14117657  0.07069812  0.030124554
#> [10,] 0.0052983203 0.12271506 0.003288349 0.12692818  0.13274165  0.280847603
#> [11,] 0.0204861857 0.06054856 0.010108004 0.09774511  0.26623776  0.097564109
#> [12,] 0.0007893835 0.11053886 0.004047914 0.15535556  0.16542508  0.204138252
#> [13,] 0.0063086632 0.10093261 0.006540056 0.11209640  0.12432007  0.199408136
#> [14,] 0.0026052914 0.07375987 0.003388709 0.07231753  0.11036575  0.171637174
#> [15,] 0.0107056476 0.10546126 0.013088392 0.22355406  0.22035064 -0.011510284
#> [16,] 0.0044313268 0.11279001 0.005406119 0.12278223  0.11588152  0.147153964
#> [17,] 0.0058698053 0.04086570 0.002166313 0.05071199 -0.01380863  0.062652779
#> [18,] 0.0033489787 0.16517970 0.013847470 0.13176515  0.11587355  0.008323070
#> [19,] 0.0070028505 0.09757772 0.004034699 0.08637714  0.01214869  0.032398720
#> [20,] 0.0034724353 0.06037453 0.003810631 0.07964989  0.08158434  0.102531239
#> [21,] 0.0062385172 0.12214671 0.005668797 0.08119527  0.01500612 -0.076470384
#> [22,] 0.0058492799 0.08573030 0.005609855 0.07959747  0.10485531  0.076300730
#> [23,] 0.0067399582 0.02526061 0.004538968 0.02172779  0.01237674  0.009652328
#> [24,] 0.0098084672 0.11990323 0.006096588 0.13064320  0.11156901  0.100776920
#> [25,] 0.0050700088 0.09397529 0.004398973 0.12859635  0.12140522  0.156622813
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768