Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12       V13       V14
#>  [1,] 0.039736611 0.28189121 0.31343001 0.33346093 0.2763446 0.1897177
#>  [2,] 0.023063094 0.22068496 0.26191465 0.28916396 0.2665790 0.2443332
#>  [3,] 0.040715097 0.41613523 0.34365790 0.27142560 0.2307320 0.2118341
#>  [4,] 0.039854512 0.25963871 0.29959953 0.37048053 0.3727367 0.3619307
#>  [5,] 0.019425130 0.15448792 0.17298899 0.19724809 0.2547273 0.3273511
#>  [6,] 0.029225310 0.22646920 0.24900795 0.25893558 0.2854310 0.3628115
#>  [7,] 0.038549406 0.24269899 0.28758060 0.27282107 0.3656127 0.5118795
#>  [8,] 0.027564007 0.38818399 0.42011145 0.42878158 0.5153129 0.4426238
#>  [9,] 0.043112020 0.32011084 0.43135288 0.49854520 0.5426374 0.5553509
#> [10,] 0.023888677 0.14425962 0.21284936 0.29076636 0.3638196 0.4418881
#> [11,] 0.033069658 0.31661090 0.34219106 0.32714883 0.2910808 0.1668303
#> [12,] 0.032776494 0.43836407 0.50132257 0.50808659 0.4775648 0.3876879
#> [13,] 0.018742797 0.21060612 0.27406230 0.30409526 0.2692734 0.1799335
#> [14,] 0.038559738 0.21353853 0.26125745 0.37219681 0.4167140 0.4956393
#> [15,] 0.023691939 0.27599701 0.34832997 0.39768761 0.4339570 0.5143249
#> [16,] 0.007834558 0.10201023 0.10296725 0.11117237 0.1401072 0.1370649
#> [17,] 0.005160174 0.04321358 0.06485408 0.10072763 0.1382145 0.1586649
#> [18,] 0.012906015 0.05936752 0.02333250 0.14301659 0.1282044 0.1208303
#> [19,] 0.015355930 0.19694607 0.20896159 0.28436557 0.4295387 0.4190781
#> [20,] 0.030835514 0.33702590 0.28639092 0.06789438 0.2407135 0.5229183
#> [21,] 0.018358653 0.25515658 0.25833081 0.11264578 0.2135448 0.4414517
#> [22,] 0.017888951 0.10737078 0.12382031 0.11141084 0.1307646 0.1056964
#> [23,] 0.025618621 0.10957790 0.09232130 0.07479569 0.0440603 0.1144991
#> [24,] 0.025951648 0.08819963 0.05964336 0.06497574 0.1696642 0.3209324
#> [25,] 0.011852467 0.10646476 0.09835006 0.12122627 0.1085831 0.1247472
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.10611797 0.1609230 0.2031220 0.2159429 0.3004153 0.030534573 0.3797379
#>  [2,] 0.20215508 0.2221085 0.2111354 0.3474743 0.5199030 0.048685446 0.6567977
#>  [3,] 0.21373206 0.2786877 0.2091861 0.2266044 0.2501294 0.038429036 0.3648966
#>  [4,] 0.30021751 0.3428091 0.3120781 0.2369039 0.2992195 0.052589410 0.4770050
#>  [5,] 0.39172199 0.5098143 0.5771409 0.6692890 0.7704939 0.007792386 0.8506076
#>  [6,] 0.43748812 0.4959045 0.6336588 0.6988423 0.7162959 0.020384383 0.7367024
#>  [7,] 0.65229718 0.7411549 0.7835239 0.8217927 0.9033173 0.048425909 0.8894422
#>  [8,] 0.28600003 0.3190734 0.3567491 0.3339836 0.5255862 0.039291525 0.7358061
#>  [9,] 0.63605990 0.7640046 0.8828546 0.9241220 0.9117604 0.067192669 0.8560775
#> [10,] 0.49740549 0.5366727 0.6674681 0.7744559 0.8663497 0.044130629 0.9141143
#> [11,] 0.16932423 0.2037455 0.2084593 0.2620340 0.2486419 0.043651815 0.1923648
#> [12,] 0.32055191 0.3194695 0.3135660 0.3031738 0.4262608 0.030359983 0.6020532
#> [13,] 0.07836762 0.2540130 0.4231003 0.6069671 0.7480362 0.007377570 0.7450834
#> [14,] 0.67171158 0.7586703 0.9408157 0.9839022 0.9058111 0.041639294 0.8350987
#> [15,] 0.66476945 0.7700816 0.8850604 0.9066718 0.8765940 0.036263667 0.7866952
#> [16,] 0.18153806 0.2012850 0.1913868 0.2106898 0.1641634 0.011329559 0.1586457
#> [17,] 0.11023148 0.1429237 0.1721492 0.1728081 0.1413711 0.007046034 0.1360898
#> [18,] 0.11979400 0.1164634 0.1218319 0.1659649 0.1435893 0.026370597 0.1667123
#> [19,] 0.44155492 0.5088165 0.4281976 0.3109091 0.3175923 0.009674253 0.5428982
#> [20,] 0.64319849 0.7273358 0.8111233 0.6693087 0.4415194 0.115097401 0.3538628
#> [21,] 0.54229156 0.7923236 0.9042401 0.9121232 0.7342161 0.027589982 0.4763224
#> [22,] 0.07170770 0.1044303 0.1471703 0.2280406 0.3876329 0.011751456 0.3801399
#> [23,] 0.08418327 0.1821357 0.1588239 0.3011104 0.3658711 0.032660910 0.3736312
#> [24,] 0.47846853 0.6471592 0.7420869 0.7754406 0.8105264 0.021828198 0.8324168
#> [25,] 0.10480809 0.1535645 0.2702985 0.3396134 0.3195457 0.025356982 0.4346868
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.4943258 0.6242593 0.7053604 0.7650096 0.8242609 0.8894063 0.9434102
#>  [2,] 0.8357903 0.8890110 0.8047450 0.8870801 0.9524336 0.9577427 0.9614229
#>  [3,] 0.3082104 0.2174045 0.2845746 0.3885180 0.5381652 0.6701273 0.7206665
#>  [4,] 0.6030960 0.5715784 0.5498989 0.6368589 0.5745881 0.7142439 0.8407795
#>  [5,] 0.9065530 0.9571641 0.9674492 0.9191825 0.8763866 0.8305932 0.8189810
#>  [6,] 0.7955035 0.8263429 0.8731923 0.9266659 0.9522025 0.9728396 0.9441924
#>  [7,] 0.9015207 0.9303441 0.8705446 0.7153358 0.5618251 0.5453742 0.4577806
#>  [8,] 0.7484409 0.7330228 0.6757953 0.6574350 0.6952483 0.8364466 0.9322890
#>  [9,] 0.7436469 0.5375515 0.3560877 0.2700806 0.2674554 0.2411298 0.2223586
#> [10,] 0.9638151 0.9817726 0.8905467 0.8456399 0.8135515 0.7584006 0.6966864
#> [11,] 0.2998266 0.6256938 0.8295883 0.8641428 0.8700705 0.8577979 0.8760906
#> [12,] 0.6939595 0.8373125 0.8749728 0.8846533 0.8549825 0.8952110 0.8717067
#> [13,] 0.7899738 0.8422417 0.9369989 0.9389542 0.9378615 0.7775896 0.4775018
#> [14,] 0.7666888 0.8298984 0.6856985 0.6098345 0.6486034 0.7288049 0.6475896
#> [15,] 0.7129575 0.6982801 0.5902063 0.4936971 0.5500201 0.6365443 0.5926648
#> [16,] 0.2332677 0.2236215 0.3582486 0.6058343 0.8576373 0.9842074 0.8364747
#> [17,] 0.2155792 0.3796786 0.4474833 0.5700169 0.6697303 0.7550613 0.7755366
#> [18,] 0.2178680 0.1720291 0.1870247 0.2310414 0.1848068 0.2824660 0.4082392
#> [19,] 0.5113619 0.4505807 0.5028322 0.6167793 0.7235718 0.7494141 0.6863899
#> [20,] 0.5225798 0.7409480 0.8360601 1.0000000 0.8608874 0.7042244 0.7135621
#> [21,] 0.5366213 0.6641990 0.7416996 0.7879501 0.7268354 0.6737905 0.6273751
#> [22,] 0.3004102 0.3319585 0.4480994 0.5895054 0.6088890 0.6127995 0.5946772
#> [23,] 0.3076313 0.1571620 0.1426609 0.2658353 0.1241911 0.2414232 0.4364463
#> [24,] 0.9668609 0.9639931 0.9567455 0.9574960 0.9143952 0.9282586 0.8504833
#> [25,] 0.5626590 0.7398771 0.8691181 0.8772085 0.8563695 0.8765756 0.8430105
#>             V28       V29           V3       V30       V31       V32       V33
#>  [1,] 0.9159197 0.8549252 0.0335025233 0.7765532 0.6759228 0.4370572 0.3588379
#>  [2,] 0.9087928 0.7876604 0.0375449148 0.5811108 0.3867257 0.3220508 0.2723307
#>  [3,] 0.8616624 0.8512195 0.0269158986 0.9167600 0.8194924 0.7434970 0.6635631
#>  [4,] 0.8803230 0.8281380 0.0615371419 0.7705100 0.7172605 0.7291356 0.5296059
#>  [5,] 0.7927186 0.6949712 0.0233891213 0.5460365 0.3367209 0.2506798 0.2125720
#>  [6,] 0.8441484 0.6917025 0.0322832540 0.5549967 0.3961221 0.2997392 0.2139656
#>  [7,] 0.4018162 0.4608211 0.0476959158 0.4406821 0.3283380 0.2614106 0.2598970
#>  [8,] 0.9022385 0.7586649 0.0529261316 0.6067731 0.5422898 0.6331059 0.5535875
#>  [9,] 0.1762139 0.1278484 0.0888222580 0.2227691 0.2470275 0.2691608 0.2970820
#> [10,] 0.5987781 0.4823499 0.0586487107 0.3692347 0.2628708 0.1821527 0.1746841
#> [11,] 0.9409123 0.8098340 0.0570992853 0.5803788 0.6468735 0.6222644 0.4835295
#> [12,] 0.6666289 0.4560409 0.0267873427 0.3967915 0.3956783 0.3872961 0.4237268
#> [13,] 0.3819108 0.3500582 0.0127661027 0.5403192 0.7117579 0.8165060 0.9082681
#> [14,] 0.6795057 0.9156974 0.0464098319 0.8705075 0.5493544 0.2477417 0.3054036
#> [15,] 0.5298691 0.5509872 0.0436582034 0.6258524 0.5985353 0.4587216 0.3639094
#> [16,] 0.6550764 0.5385626 0.0118599373 0.3865404 0.3163826 0.1947716 0.1902780
#> [17,] 0.8517391 0.9297059 0.0129231712 0.8987238 0.7424043 0.6736118 0.7038664
#> [18,] 0.5111253 0.5623005 0.0282209416 0.6188459 0.7092000 0.6622542 0.6890717
#> [19,] 0.7026261 0.6277210 0.0007962256 0.5801847 0.5588608 0.5500234 0.4838275
#> [20,] 0.5372667 0.4279239 0.1345625466 0.4814556 0.4258340 0.1118098 0.1540169
#> [21,] 0.4338110 0.2726166 0.0454311346 0.3234903 0.3080099 0.3554062 0.2958747
#> [22,] 0.7683268 0.9806285 0.0158460636 0.9752485 0.9262513 0.8649432 0.8046823
#> [23,] 0.7198145 0.8715149 0.0552350424 0.7683321 0.4923560 0.2336293 0.5165947
#> [24,] 0.6610330 0.4305366 0.0376114219 0.4395509 0.4409626 0.3379315 0.2788595
#> [25,] 0.7882752 0.8096025 0.0303671007 0.7928207 0.8002061 0.7605314 0.7461390
#>              V34        V35        V36        V37        V38        V39
#>  [1,] 0.32207090 0.20891407 0.07417064 0.15504914 0.11517387 0.11538574
#>  [2,] 0.25042862 0.21941895 0.11186416 0.11950323 0.17183014 0.16105631
#>  [3,] 0.62542801 0.66174619 0.68618823 0.73196204 0.68958042 0.61490795
#>  [4,] 0.33308231 0.20884949 0.21780548 0.25947124 0.36815041 0.42628783
#>  [5,] 0.14387047 0.21623109 0.23749342 0.24263114 0.24495148 0.27258339
#>  [6,] 0.10012403 0.09337479 0.12930875 0.22483947 0.25114133 0.21445220
#>  [7,] 0.22630887 0.19256821 0.18813373 0.19546585 0.21377563 0.27655236
#>  [8,] 0.35641324 0.11963856 0.22303651 0.25217566 0.35734030 0.56842547
#>  [9,] 0.33008552 0.36464343 0.19852612 0.26918519 0.31536367 0.20976738
#> [10,] 0.17007430 0.21421553 0.28891016 0.29727906 0.34188449 0.38670240
#> [11,] 0.20598030 0.08666001 0.10875106 0.20164372 0.18505608 0.18585375
#> [12,] 0.38317461 0.18782961 0.31660820 0.40778253 0.32963079 0.32960989
#> [13,] 0.85697950 0.64995469 0.43590111 0.14377849 0.19608362 0.37938565
#> [14,] 0.43199135 0.29914385 0.12322962 0.20619348 0.18955537 0.09124201
#> [15,] 0.30902669 0.21404256 0.23877778 0.32660674 0.26817982 0.25689089
#> [16,] 0.20749903 0.28377283 0.22085153 0.07774765 0.07227826 0.12989434
#> [17,] 0.60052519 0.54461261 0.47129804 0.36641561 0.34886229 0.33412150
#> [18,] 0.82764462 0.92779491 0.90030375 0.75253509 0.62521319 0.58584867
#> [19,] 0.50488144 0.55000583 0.54578449 0.38965436 0.24331648 0.34330508
#> [20,] 0.05667899 0.38522062 0.68652548 0.69988567 0.33142730 0.08572524
#> [21,] 0.22626870 0.48398281 0.65142966 0.57235267 0.35985882 0.21929613
#> [22,] 0.71547054 0.60329283 0.49880064 0.44036550 0.22533833 0.09496134
#> [23,] 0.57833106 0.32585438 0.42192231 0.58526873 0.62377076 0.58064440
#> [24,] 0.28275958 0.19066927 0.08551635 0.14946433 0.12394843 0.08027447
#> [25,] 0.73041154 0.69179052 0.58954227 0.48511939 0.36199760 0.24209392
#>                V4       V40       V41        V42        V43        V44
#>  [1,]  0.04766526 0.1930279 0.2613173 0.29115138 0.25942211 0.13589147
#>  [2,]  0.03121256 0.1784753 0.1325693 0.14111441 0.19446281 0.20526946
#>  [3,]  0.06567961 0.5311461 0.4241812 0.35985016 0.38229894 0.47923230
#>  [4,]  0.02049258 0.3791459 0.5593160 0.68457753 0.58908110 0.51502904
#>  [5,]  0.07129592 0.3164917 0.3474614 0.33130965 0.24961105 0.20787660
#>  [6,]  0.04028409 0.2620276 0.2408370 0.20492859 0.20332708 0.15644736
#>  [7,]  0.06712249 0.2498940 0.1722382 0.17447121 0.22236528 0.17811565
#>  [8,]  0.04728608 0.4465648 0.2769279 0.40193023 0.40844879 0.45458936
#>  [9,]  0.09999094 0.2010964 0.2997813 0.26419715 0.15516154 0.10868229
#> [10,]  0.06427188 0.3140033 0.2636411 0.28071893 0.19687967 0.17743719
#> [11,]  0.08104031 0.1188763 0.1769687 0.11684662 0.11915782 0.13649298
#> [12,]  0.02648710 0.3000670 0.1006707 0.23524949 0.35994750 0.40538082
#> [13,]  0.03944151 0.5314656 0.5125308 0.36352762 0.22562737 0.10790937
#> [14,]  0.01418223 0.1694754 0.1537447 0.16757191 0.14550056 0.09434508
#> [15,]  0.04985390 0.3086230 0.2509249 0.23480387 0.12368200 0.12010773
#> [16,]  0.02904756 0.2009050 0.2579574 0.24597621 0.18973905 0.10033289
#> [17,]  0.02605273 0.2822715 0.1807503 0.07990950 0.11792241 0.13212330
#> [18,]  0.02979871 0.5708332 0.4901885 0.43177949 0.29151305 0.20864397
#> [19,] -0.01290350 0.4762382 0.3422967 0.26994200 0.19263061 0.14724128
#> [20,]  0.11435853 0.3322975 0.4187536 0.39013304 0.23950903 0.19239246
#> [21,]  0.05460849 0.3466762 0.3853218 0.32020387 0.20385031 0.19506449
#> [22,]  0.02721025 0.1655668 0.1911006 0.21869709 0.27723252 0.35018279
#> [23,]  0.06152972 0.3825513 0.2300429 0.10836647 0.04852122 0.02122101
#> [24,]  0.04885776 0.1435928 0.1066935 0.05838325 0.04656984 0.11099644
#> [25,]  0.03390419 0.2775987 0.2230409 0.14910890 0.10080178 0.03622010
#>               V45         V46        V47         V48         V49         V5
#>  [1,] 0.122437409 0.134244545 0.14231289 0.114775141 0.059575697 0.06196221
#>  [2,] 0.207350794 0.241973537 0.18464177 0.099680166 0.060258163 0.05249726
#>  [3,] 0.494822811 0.345194942 0.20074541 0.189449547 0.100940590 0.07735701
#>  [4,] 0.589837216 0.469957159 0.31575392 0.231327170 0.129498863 0.09247124
#>  [5,] 0.152859530 0.121356051 0.12205076 0.080140525 0.052082918 0.10409756
#>  [6,] 0.119826473 0.071563510 0.10819422 0.082891879 0.050400581 0.10668505
#>  [7,] 0.112431383 0.096474704 0.08329060 0.066548849 0.048528333 0.07856555
#>  [8,] 0.498444940 0.260530390 0.15781247 0.183574243 0.101091032 0.10098500
#>  [9,] 0.076277093 0.058706657 0.07090513 0.058646447 0.035102918 0.10096848
#> [10,] 0.149132820 0.112760508 0.09263405 0.082234080 0.060659477 0.14744823
#> [11,] 0.079211221 0.074242830 0.10902648 0.135684803 0.082457499 0.10258203
#> [12,] 0.377611785 0.205783691 0.13002367 0.085229747 0.032590855 0.04250504
#> [13,] 0.087184817 0.114503174 0.12226044 0.108676728 0.078980693 0.05284297
#> [14,] 0.205190748 0.232002920 0.16933938 0.098040235 0.058528813 0.01554706
#> [15,] 0.151215898 0.152118638 0.15252228 0.108423034 0.050009298 0.08595995
#> [16,] 0.092365441 0.096203230 0.04222917 0.039184855 0.029910969 0.04952552
#> [17,] 0.122920947 0.057931291 0.02862489 0.036860354 0.017941848 0.02732703
#> [18,] 0.165655367 0.116506135 0.11579675 0.103044487 0.048847704 0.02809867
#> [19,] 0.061738157 0.086001612 0.03415904 0.008913281 0.008885864 0.04536349
#> [20,] 0.219665009 0.178242726 0.16806682 0.172502654 0.106769124 0.08003726
#> [21,] 0.156964096 0.070690150 0.09888194 0.106611280 0.060007053 0.11570412
#> [22,] 0.327809676 0.217259040 0.10280239 0.042180527 0.027981685 0.03028350
#> [23,] 0.008452256 0.003669196 0.02330308 0.024133511 0.007225990 0.04918187
#> [24,] 0.098918611 0.037436567 0.07411228 0.054265872 0.023113112 0.07484376
#> [25,] 0.007100059 0.030084020 0.02877779 0.038555612 0.009800790 0.07339824
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.006570489 0.014063357 0.013630334 0.008052137 0.004386723 0.005417487
#>  [2,] 0.022422524 0.014730347 0.006834503 0.004497309 0.009449799 0.005677896
#>  [3,] 0.030615782 0.019127562 0.020818944 0.012914881 0.010404073 0.008554623
#>  [4,] 0.045447308 0.040414298 0.022916207 0.011251454 0.013421959 0.010861534
#>  [5,] 0.022944836 0.014131869 0.013863100 0.014180473 0.009394837 0.004709613
#>  [6,] 0.025514382 0.017982248 0.015691259 0.018047080 0.009487538 0.003743491
#>  [7,] 0.018374890 0.022399366 0.010526209 0.006291957 0.015217576 0.009068268
#>  [8,] 0.030208173 0.030698400 0.018542800 0.008967262 0.011088781 0.017159038
#>  [9,] 0.021508798 0.018939225 0.027243226 0.020688575 0.012083816 0.020856109
#> [10,] 0.020658116 0.018525613 0.016848423 0.012704084 0.023429352 0.013153374
#> [11,] 0.025137341 0.016247093 0.013325937 0.008260168 0.010702032 0.005799599
#> [12,] 0.009079821 0.007963574 0.008256437 0.012495530 0.010818475 0.004109003
#> [13,] 0.024810239 0.014449257 0.008408698 0.004619791 0.002141893 0.006751929
#> [14,] 0.038990928 0.011679773 0.002731991 0.002987708 0.011899327 0.013956349
#> [15,] 0.014862431 0.009539593 0.009544727 0.006918913 0.008068932 0.009772886
#> [16,] 0.008405211 0.007444795 0.011578112 0.005231725 0.004757720 0.005291197
#> [17,] 0.007355763 0.006657148 0.006943643 0.011524195 0.006181784 0.006263722
#> [18,] 0.018976558 0.018217829 0.005268016 0.010140974 0.009606986 0.010055821
#> [19,] 0.014008773 0.017961818 0.006265489 0.006170286 0.015023291 0.012028886
#> [20,] 0.020146489 0.021188453 0.010185544 0.005770427 0.016765256 0.012846021
#> [21,] 0.012949985 0.024055404 0.016974858 0.009558985 0.011590786 0.007022440
#> [22,] 0.016798727 0.010991608 0.005390506 0.005864630 0.010274917 0.002462790
#> [23,] 0.003651320 0.001943758 0.010285166 0.024583374 0.018968834 0.010416365
#> [24,] 0.014040328 0.008047234 0.011726142 0.010223687 0.006209097 0.004061928
#> [25,] 0.012175282 0.009083585 0.007941554 0.008586563 0.007676138 0.007681940
#>               V56         V57         V58         V59         V6         V60
#>  [1,] 0.003903102 0.004460521 0.004001016 0.003003980 0.09408125 0.004806726
#>  [2,] 0.003819858 0.003096883 0.002391141 0.003529009 0.03899035 0.004799695
#>  [3,] 0.007289221 0.008322918 0.004925341 0.002675173 0.09125793 0.003845506
#>  [4,] 0.009106199 0.005913646 0.007938962 0.006119443 0.12899904 0.004111776
#>  [5,] 0.010260606 0.007679008 0.006261012 0.007805336 0.09416883 0.006204688
#>  [6,] 0.008870413 0.009520206 0.010480599 0.008775331 0.14924506 0.007817318
#>  [7,] 0.009643623 0.008535135 0.011055827 0.008733653 0.13019349 0.007087478
#>  [8,] 0.010259971 0.009369505 0.011184097 0.007740913 0.12494507 0.005247295
#>  [9,] 0.014851594 0.007516677 0.015689981 0.014826803 0.08912893 0.010704544
#> [10,] 0.007272403 0.008633786 0.008784017 0.009877635 0.18254139 0.006332516
#> [11,] 0.010013055 0.006726931 0.007250734 0.009587457 0.13445048 0.009636772
#> [12,] 0.005069842 0.001632017 0.004133845 0.007661164 0.08405407 0.003597922
#> [13,] 0.004930822 0.002542233 0.007113052 0.006562000 0.09280052 0.003255134
#> [14,] 0.008294482 0.005521647 0.004893348 0.004343851 0.10661587 0.010880560
#> [15,] 0.009793340 0.011508331 0.006799768 0.004935495 0.11611994 0.004795601
#> [16,] 0.001944695 0.002454865 0.003767721 0.002448693 0.08889023 0.002141772
#> [17,] 0.007201373 0.006047038 0.001926788 0.005184538 0.03472165 0.003993798
#> [18,] 0.009232428 0.006975197 0.008392478 0.011053313 0.06424314 0.005793345
#> [19,] 0.011587115 0.009659710 0.005079575 0.011328686 0.16083048 0.005248632
#> [20,] 0.005575270 0.009025891 0.008237792 0.010193335 0.10682123 0.006403834
#> [21,] 0.004583166 0.007176508 0.012135185 0.017591288 0.13037306 0.013113331
#> [22,] 0.001905382 0.004139405 0.003721923 0.006523277 0.03750309 0.002360473
#> [23,] 0.010853611 0.012155757 0.007172764 0.012507393 0.08907450 0.010443913
#> [24,] 0.012871056 0.009957269 0.002288624 0.005772632 0.08091280 0.004228046
#> [25,] 0.007928251 0.009719533 0.006924564 0.004585171 0.10861971 0.008007590
#>               V7         V8         V9
#>  [1,] 0.09515602 0.15352516 0.20752049
#>  [2,] 0.09669427 0.10173479 0.17574989
#>  [3,] 0.13280201 0.19784216 0.37848564
#>  [4,] 0.12919770 0.16190742 0.26156821
#>  [5,] 0.13392538 0.11385129 0.12368213
#>  [6,] 0.14491403 0.16245777 0.21461198
#>  [7,] 0.18177472 0.21342277 0.24597121
#>  [8,] 0.12076152 0.14478554 0.26277980
#>  [9,] 0.10261853 0.08051430 0.17792761
#> [10,] 0.15284642 0.13991469 0.15159082
#> [11,] 0.15096037 0.19564821 0.23245163
#> [12,] 0.11632264 0.13239051 0.32471358
#> [13,] 0.09613732 0.16676582 0.23075186
#> [14,] 0.12753918 0.10500957 0.19782097
#> [15,] 0.14194320 0.17470387 0.24931163
#> [16,] 0.09323978 0.07877201 0.11109593
#> [17,] 0.03391205 0.02308908 0.03941473
#> [18,] 0.05941731 0.04943350 0.01220292
#> [19,] 0.18485161 0.19228108 0.16922603
#> [20,] 0.17186130 0.21642175 0.27653971
#> [21,] 0.16454443 0.17377187 0.18624952
#> [22,] 0.08461192 0.07139679 0.07816388
#> [23,] 0.02010905 0.06861787 0.13259275
#> [24,] 0.11371232 0.10343768 0.11071673
#> [25,] 0.10862887 0.11509084 0.10214486
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623