Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13       V14
#>  [1,] 0.02928677 0.30671446 0.36078259 0.38022926 0.42786789 0.3873948
#>  [2,] 0.02380068 0.21058830 0.25987664 0.29725714 0.26951201 0.2368689
#>  [3,] 0.03958928 0.27691733 0.38243340 0.53686097 0.57947671 0.6801495
#>  [4,] 0.03700907 0.18118326 0.26200247 0.32433057 0.41934622 0.5820619
#>  [5,] 0.06305593 0.52817378 0.46564248 0.30846115 0.21488070 0.2791704
#>  [6,] 0.02279156 0.33293526 0.25967849 0.30467727 0.25553332 0.2031458
#>  [7,] 0.10187256 0.16592669 0.27062926 0.23270262 0.27434857 0.3398807
#>  [8,] 0.03140317 0.23915962 0.30480522 0.32263259 0.26597983 0.1306890
#>  [9,] 0.01846900 0.60739208 0.65703703 0.47217912 0.37992599 0.2922834
#> [10,] 0.01613713 0.17103910 0.11665352 0.06667616 0.16566638 0.2695961
#> [11,] 0.06349730 0.33871745 0.51444699 0.50665718 0.50918503 0.5192406
#> [12,] 0.02966493 0.16890465 0.24441896 0.27913510 0.25939319 0.2287649
#> [13,] 0.03865090 0.42237177 0.50348675 0.53862334 0.52978519 0.4169083
#> [14,] 0.01571249 0.09444505 0.10956337 0.10375324 0.12379480 0.1606513
#> [15,] 0.01551534 0.14297776 0.12050041 0.06775598 0.07162609 0.1024410
#> [16,] 0.02180244 0.10765189 0.06157588 0.01437792 0.13302310 0.2978494
#> [17,] 0.01335100 0.15664717 0.14264599 0.17990773 0.31463270 0.2721097
#> [18,] 0.01346228 0.08492048 0.11736002 0.14866835 0.20062812 0.2145210
#> [19,] 0.00980733 0.02407156 0.10269960 0.13738965 0.23164044 0.3179437
#> [20,] 0.01295030 0.12070254 0.16742440 0.17529464 0.20862799 0.2718754
#> [21,] 0.03474670 0.23358213 0.25575890 0.17210297 0.26656211 0.4018675
#> [22,] 0.01636623 0.08122508 0.08266550 0.13812622 0.12563199 0.1269295
#> [23,] 0.01612162 0.13312460 0.12122752 0.15098238 0.14947186 0.2297293
#> [24,] 0.02314399 0.61947068 0.63990195 0.72664417 0.57334924 0.5396739
#> [25,] 0.01368337 0.07334285 0.10914169 0.11853730 0.20150127 0.2607555
#>             V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.3550417 0.37947824 0.3582778 0.2889631 0.3795485 0.048859720 0.6106452
#>  [2,] 0.1955755 0.23009311 0.2721341 0.3900465 0.5814145 0.037044407 0.6800305
#>  [3,] 0.7195399 0.82454037 0.9520337 0.9111339 0.9419821 0.043500244 0.8972072
#>  [4,] 0.6659866 0.64017162 0.6693019 0.7824783 0.8967797 0.052304337 0.9199980
#>  [5,] 0.3514847 0.39777446 0.2222044 0.2381260 0.3190176 0.082663085 0.5898587
#>  [6,] 0.1213613 0.14413341 0.1304510 0.1880306 0.1494803 0.028065402 0.1351532
#>  [7,] 0.3883509 0.43686316 0.4398420 0.4405211 0.4817537 0.184445206 0.3427131
#>  [8,] 0.1277903 0.17511770 0.2058430 0.2505774 0.2463177 0.018426962 0.2953634
#>  [9,] 0.1727952 0.08993279 0.1925282 0.3960551 0.6064410 0.024326405 0.7273784
#> [10,] 0.3639906 0.50022475 0.5568450 0.6378203 0.7762733 0.006470339 0.8427596
#> [11,] 0.6305984 0.83764630 0.9657818 0.9305687 0.8094880 0.111669323 0.7530562
#> [12,] 0.2120355 0.19229106 0.2409149 0.2659014 0.3129405 0.037101893 0.3289276
#> [13,] 0.2758804 0.24717934 0.2849578 0.3090709 0.5053814 0.040230310 0.7006719
#> [14,] 0.1992687 0.19663132 0.1717859 0.1674210 0.1076993 0.015484215 0.1318452
#> [15,] 0.1145878 0.15281746 0.2102093 0.2634645 0.3266083 0.010874207 0.4967536
#> [16,] 0.5364358 0.76881274 0.8335999 0.8096259 0.8694821 0.018201605 0.8713060
#> [17,] 0.1950123 0.33637363 0.4064528 0.4448558 0.6660032 0.045759218 0.7415744
#> [18,] 0.2519410 0.27301278 0.2491692 0.2098268 0.4239389 0.012838927 0.5345850
#> [19,] 0.4386459 0.49035362 0.5480381 0.6309102 0.7070152 0.008848059 0.7978121
#> [20,] 0.1948599 0.20517138 0.2654479 0.3578555 0.4231248 0.008446184 0.4569971
#> [21,] 0.5109487 0.60494395 0.6261125 0.6193603 0.5810037 0.058473494 0.5487964
#> [22,] 0.1380397 0.17812420 0.2003705 0.2375958 0.2870422 0.026269072 0.2858394
#> [23,] 0.2343684 0.30847954 0.3690079 0.4143896 0.4085709 0.024256001 0.4325409
#> [24,] 0.6562824 0.70349739 0.6956206 0.7782731 0.9138371 0.056333083 0.8601480
#> [25,] 0.2712059 0.19627201 0.2328484 0.3120575 0.3990776 0.006278989 0.4669383
#>              V21       V22       V23        V24       V25       V26       V27
#>  [1,] 0.71217470 0.6439584 0.5564499 0.60560298 0.6403096 0.8109011 0.8985433
#>  [2,] 0.79265243 0.8163480 0.8244835 0.92141270 0.9305984 0.9665016 0.9847409
#>  [3,] 0.83068548 0.7823808 0.6397361 0.40494304 0.2569112 0.1344764 0.2021190
#>  [4,] 0.90758596 0.9083297 0.9536382 0.90838296 0.6986148 0.5667855 0.5012553
#>  [5,] 0.53655229 0.2714055 0.1994490 0.25937349 0.4220730 0.6406033 0.6297775
#>  [6,] 0.03181473 0.1778377 0.4278971 0.58195900 0.6947352 0.7997199 0.9288414
#>  [7,] 0.22516353 0.1701844 0.3826399 0.61792031 0.4750270 0.3156145 0.4439202
#>  [8,] 0.40791503 0.6251637 0.7883334 0.85002663 0.9481677 0.9885960 0.9702540
#>  [9,] 0.80348191 0.8654017 0.8845475 0.93310154 0.8471699 0.7663530 0.7326929
#> [10,] 0.91147897 0.9579209 0.9730015 0.81975702 0.6464689 0.6281867 0.6630899
#> [11,] 0.71806081 0.4100551 0.0913522 0.09047487 0.2660735 0.2496806 0.2015928
#> [12,] 0.44409062 0.5184341 0.6182807 0.73379140 0.7831827 0.7918737 0.8539588
#> [13,] 0.82202280 0.8471514 0.7529032 0.74350359 0.7357206 0.8192130 0.9079401
#> [14,] 0.20694830 0.2219831 0.3484641 0.58153754 0.8346294 0.9739297 0.8472748
#> [15,] 0.61837062 0.7698890 0.8917378 0.84338360 0.8314330 0.8988748 0.8937034
#> [16,] 0.95632689 0.9553127 0.9485244 0.94602797 0.9013793 0.9419321 0.9196505
#> [17,] 0.69398628 0.5539263 0.4899494 0.47363834 0.7060411 0.8312229 0.7894222
#> [18,] 0.57238803 0.6152827 0.7315046 0.89675687 0.9301797 0.8639445 0.9220419
#> [19,] 0.96901029 1.0152041 0.9660184 0.93935610 0.8550226 0.7853592 0.6519447
#> [20,] 0.54071887 0.5779151 0.5611869 0.68918922 0.9417291 0.9741734 1.0072306
#> [21,] 0.53692720 0.5401131 0.7027012 0.83441089 0.7174808 0.6731596 0.7537103
#> [22,] 0.30410192 0.2438039 0.2956106 0.36722254 0.3268155 0.3823848 0.4827643
#> [23,] 0.48682613 0.5512917 0.6332474 0.65854700 0.6738038 0.6703399 0.5640622
#> [24,] 0.78999326 0.6847447 0.4432960 0.37699629 0.5456963 0.2305269 0.5164576
#> [25,] 0.47979756 0.5659180 0.6267680 0.67633757 0.8509736 0.9035634 0.9504139
#>             V28        V29          V3       V30       V31       V32       V33
#>  [1,] 0.8872623 0.83105837 0.073918954 0.7151053 0.6755955 0.7209581 0.5520834
#>  [2,] 0.9670771 0.85633022 0.038165763 0.6400744 0.4341732 0.3102400 0.2571460
#>  [3,] 0.1816213 0.13612226 0.032999035 0.1354807 0.1003902 0.1647297 0.2453040
#>  [4,] 0.4241984 0.40700737 0.070430352 0.3279406 0.2200869 0.1485648 0.1570769
#>  [5,] 0.7709918 0.75100121 0.045075654 0.9051016 0.6717085 0.6624276 0.6926472
#>  [6,] 0.9684101 0.89702887 0.019305288 0.9144867 0.9289991 0.8974051 0.7656466
#>  [7,] 0.5652237 0.50164383 0.241916785 0.4190429 0.3732364 0.2749024 0.1865735
#>  [8,] 0.9631883 0.86888608 0.028382863 0.7485158 0.7195044 0.5867813 0.4373686
#>  [9,] 0.7225824 0.58034829 0.043399259 0.4409233 0.4113600 0.4242887 0.6638952
#> [10,] 0.7044625 0.63237831 0.026811577 0.4903068 0.3125391 0.3272104 0.2920628
#> [11,] 0.1822781 0.05930877 0.152814774 0.2269002 0.1981047 0.2001089 0.2578370
#> [12,] 0.8790256 0.80535093 0.049978478 0.6804334 0.6415455 0.5088463 0.3556638
#> [13,] 0.8656120 0.70188388 0.047563541 0.4986991 0.4031101 0.3876033 0.3595632
#> [14,] 0.6765511 0.56972028 0.018311945 0.4496635 0.3575055 0.2125628 0.1982257
#> [15,] 0.8332977 0.77843637 0.009483805 0.7084080 0.6730848 0.5880478 0.5471336
#> [16,] 0.7288455 0.44612418 0.037833007 0.3832904 0.4238909 0.3801315 0.3216121
#> [17,] 0.5096312 0.32981547 0.047582293 0.2235499 0.2149738 0.4239187 0.5225762
#> [18,] 0.8248708 0.59539745 0.017022706 0.4402241 0.3601481 0.2717050 0.2037297
#> [19,] 0.4510305 0.49816112 0.003596690 0.6453987 0.5033925 0.3143007 0.3783418
#> [20,] 0.9787381 0.86566268 0.008373836 0.7080136 0.5961444 0.4587490 0.3101991
#> [21,] 0.7334602 0.59015573 0.070449354 0.5664807 0.5232898 0.3370626 0.2838174
#> [22,] 0.6827860 0.79534453 0.022833774 0.7966588 0.7481325 0.6490050 0.6901274
#> [23,] 0.5623735 0.53973664 0.044764596 0.4933199 0.6315830 0.6248736 0.5238378
#> [24,] 0.8660436 0.60298509 0.111948812 0.8519297 0.8670325 0.5064772 0.1702141
#> [25,] 0.9321316 0.76714198 0.012948436 0.6500590 0.6392213 0.4935511 0.3527119
#>             V34        V35        V36        V37        V38        V39
#>  [1,] 0.3307676 0.14996589 0.23717841 0.32395850 0.37632018 0.48785575
#>  [2,] 0.2051524 0.17166033 0.08433197 0.06812703 0.12151615 0.12389299
#>  [3,] 0.2190950 0.26410544 0.16023237 0.28051273 0.29891226 0.20531819
#>  [4,] 0.1758851 0.24677256 0.30315816 0.29485890 0.36778328 0.35775673
#>  [5,] 0.7122991 0.79830669 0.79084000 0.79775191 0.73820172 0.58150125
#>  [6,] 0.6547619 0.64488878 0.69364683 0.72045113 0.58498843 0.50344456
#>  [7,] 0.2137956 0.27205359 0.38295592 0.29716647 0.30293639 0.36695787
#>  [8,] 0.2106370 0.10425031 0.09782094 0.12833588 0.14961816 0.16642958
#>  [9,] 0.7334535 0.58873311 0.46171073 0.41023031 0.35751606 0.42253057
#> [10,] 0.1838970 0.20313877 0.20992647 0.17618242 0.18354278 0.20667374
#> [11,] 0.2409906 0.31615991 0.15539827 0.10763684 0.20508574 0.19649685
#> [12,] 0.2818201 0.19656957 0.16802977 0.25050199 0.19609387 0.22474026
#> [13,] 0.2852270 0.10815283 0.22533330 0.25798967 0.28814165 0.44357958
#> [14,] 0.2382087 0.30224044 0.23245384 0.08618132 0.06091547 0.10690221
#> [15,] 0.5589209 0.55043575 0.49863190 0.46653406 0.33312910 0.18716916
#> [16,] 0.3146719 0.24726471 0.11164294 0.22338981 0.17093699 0.07417831
#> [17,] 0.4492757 0.39687766 0.51841389 0.60676464 0.39835104 0.54101811
#> [18,] 0.1588372 0.16354799 0.17274021 0.20860648 0.28142567 0.23808599
#> [19,] 0.3271886 0.06652572 0.02756640 0.00672927 0.02388086 0.15726763
#> [20,] 0.1646904 0.06440601 0.24891583 0.24732599 0.32131116 0.36253555
#> [21,] 0.1567329 0.38093325 0.60380304 0.57222505 0.28738896 0.19754135
#> [22,] 0.7541222 0.76126206 0.74545306 0.64426389 0.48056289 0.37898576
#> [23,] 0.5462372 0.61292840 0.60579756 0.50073732 0.47274254 0.49272294
#> [24,] 0.2509955 0.40519648 0.28290234 0.60425506 0.66708088 0.52723072
#> [25,] 0.1964001 0.11800468 0.23027422 0.24482755 0.28011656 0.29992481
#>                V4        V40        V41        V42         V43        V44
#>  [1,] 0.059692922 0.36588878 0.42346672 0.55005182  0.47823627 0.46441928
#>  [2,] 0.029085569 0.13316127 0.06248115 0.11104174  0.19763284 0.20385652
#>  [3,] 0.093873789 0.14001941 0.19861260 0.30312698  0.24107424 0.03866794
#>  [4,] 0.104571421 0.28362824 0.23688769 0.22711326  0.26939318 0.21350591
#>  [5,] 0.107276359 0.45484430 0.32888602 0.32987057  0.39965219 0.47382897
#>  [6,] 0.033614683 0.53871159 0.50910706 0.44524215  0.46308993 0.54521981
#>  [7,] 0.337984062 0.40790068 0.34422639 0.27972935  0.18982535 0.19636075
#>  [8,] 0.045539767 0.06799716 0.12750228 0.12141993  0.16214737 0.16357371
#>  [9,] 0.057201207 0.45329654 0.34663041 0.27816614  0.23751598 0.19759900
#> [10,] 0.074914033 0.21869149 0.29319414 0.31167052  0.21754351 0.21994125
#> [11,] 0.132166431 0.04217448 0.11861638 0.14912492  0.14387758 0.14729537
#> [12,] 0.059697306 0.28233898 0.31335430 0.32382931  0.22377714 0.13241842
#> [13,] 0.039733166 0.38738805 0.13708216 0.27944023  0.37718528 0.43026758
#> [14,] 0.019414020 0.17539824 0.24632093 0.23251179  0.19159214 0.10791925
#> [15,] 0.023066127 0.15553000 0.15012363 0.08523639  0.08153212 0.03417511
#> [16,] 0.036239361 0.09793907 0.06869997 0.02436348  0.05098472 0.11314394
#> [17,] 0.047497587 0.52523056 0.38186276 0.33466373  0.26624098 0.24171743
#> [18,] 0.005701433 0.12898535 0.11914773 0.17162237  0.16480962 0.15212938
#> [19,] 0.045769893 0.20187571 0.14858713 0.02824262 -0.01957769 0.06136057
#> [20,] 0.025437954 0.23582754 0.08622342 0.22614145  0.17888908 0.15907144
#> [21,] 0.069443555 0.24686502 0.31397433 0.32805001  0.23129135 0.14954449
#> [22,] 0.026409162 0.35885841 0.34383288 0.31429387  0.25886188 0.21532036
#> [23,] 0.045013917 0.50931313 0.34671424 0.27710645  0.26467236 0.26386035
#> [24,] 0.106677761 0.46861672 0.46701473 0.24965853  0.20208656 0.21369543
#> [25,] 0.004354006 0.18251874 0.12578541 0.25951110  0.23841950 0.15465799
#>               V45          V46        V47        V48        V49         V5
#>  [1,] 0.559547609 3.493860e-01 0.22487714 0.20011887 0.11434268 0.09260249
#>  [2,] 0.193206916 1.920389e-01 0.15306639 0.09546123 0.05104431 0.08219816
#>  [3,] 0.157711694 1.183936e-01 0.04346355 0.05062607 0.04433376 0.19371325
#>  [4,] 0.117610647 4.481291e-02 0.05151757 0.05650034 0.04018156 0.13053531
#>  [5,] 0.507909672 3.342380e-01 0.12071393 0.11353038 0.07142123 0.11600146
#>  [6,] 0.540669873 4.178744e-01 0.32705575 0.24386274 0.12730079 0.05681992
#>  [7,] 0.264296497 2.140432e-01 0.16529571 0.13973965 0.06659537 0.31429317
#>  [8,] 0.101182478 5.255531e-02 0.09863573 0.10974631 0.06170241 0.05546562
#>  [9,] 0.179698321 1.281705e-01 0.11720210 0.13111009 0.06475103 0.04539086
#> [10,] 0.157441086 1.283656e-01 0.10241731 0.05413839 0.03792179 0.11201114
#> [11,] 0.063198063 1.694596e-02 0.08552863 0.08431300 0.02323580 0.14792075
#> [12,] 0.136200292 1.815310e-01 0.18914057 0.14630944 0.06387902 0.07341788
#> [13,] 0.391556351 1.710921e-01 0.08718556 0.10647104 0.06708675 0.08214395
#> [14,] 0.085112755 8.162332e-02 0.03985896 0.03842139 0.03161432 0.03938964
#> [15,] 0.007458627 3.150650e-02 0.03198479 0.03306720 0.01596509 0.05929770
#> [16,] 0.114479258 3.464083e-02 0.04693656 0.03944297 0.01968916 0.06480669
#> [17,] 0.107263560 7.094289e-02 0.05792228 0.03724742 0.04408928 0.06078635
#> [18,] 0.132616826 8.412709e-02 0.06979783 0.05164125 0.01960815 0.05187801
#> [19,] 0.058345104 3.919322e-03 0.03702888 0.03037573 0.02307953 0.07518186
#> [20,] 0.125114868 8.108822e-02 0.11932244 0.10260903 0.04502934 0.01514703
#> [21,] 0.161735636 1.506776e-01 0.11227035 0.10465309 0.05864222 0.05052032
#> [22,] 0.180743679 1.242428e-01 0.08939415 0.06754089 0.04136727 0.03760770
#> [23,] 0.141361470 1.040878e-01 0.09430748 0.06277929 0.02824717 0.06907306
#> [24,] 0.197376326 5.773542e-05 0.13700465 0.07268534 0.01123880 0.09592196
#> [25,] 0.139701143 1.169108e-01 0.09535552 0.07238636 0.02738226 0.03785186
#>               V50         V51         V52         V53         V54          V55
#>  [1,] 0.032668125 0.030828031 0.016240240 0.008045122 0.010735500 0.0118117961
#>  [2,] 0.016893227 0.019826809 0.012314792 0.008556039 0.005480725 0.0049970942
#>  [3,] 0.041279802 0.021189242 0.029263977 0.021431630 0.016984780 0.0236192444
#>  [4,] 0.013791388 0.022811771 0.014338433 0.012837402 0.018292153 0.0183284750
#>  [5,] 0.025875818 0.011929906 0.025720583 0.016844883 0.016137229 0.0101018454
#>  [6,] 0.038773104 0.022597942 0.017538945 0.007380746 0.007736265 0.0068013451
#>  [7,] 0.018461034 0.013551075 0.032099918 0.020504400 0.015648512 0.0211259446
#>  [8,] 0.018391589 0.008554209 0.008085135 0.008373174 0.011636678 0.0055775737
#>  [9,] 0.025628521 0.009360361 0.011793149 0.006820850 0.017334038 0.0121023220
#> [10,] 0.011692976 0.011329729 0.015202074 0.009586068 0.006710383 0.0001886777
#> [11,] 0.009825688 0.025532351 0.013592354 0.024381114 0.010708664 0.0146764112
#> [12,] 0.015624022 0.020967104 0.016671576 0.008857450 0.010519676 0.0075069454
#> [13,] 0.013842209 0.022392122 0.014742633 0.008453253 0.007602186 0.0068088675
#> [14,] 0.010417528 0.009262311 0.011330854 0.005701792 0.004882356 0.0063401051
#> [15,] 0.007355452 0.001623697 0.002761125 0.007395672 0.007816365 0.0101114280
#> [16,] 0.008668343 0.013036650 0.011618713 0.010749899 0.006685463 0.0037364787
#> [17,] 0.020104253 0.011547555 0.011575666 0.006909867 0.019686060 0.0186660890
#> [18,] 0.007529953 0.001046615 0.005357101 0.006784513 0.008883476 0.0113217023
#> [19,] 0.008371344 0.015840927 0.001287828 0.006601993 0.011187608 0.0089175824
#> [20,] 0.014635414 0.002653032 0.004579712 0.003139083 0.003510102 0.0014114011
#> [21,] 0.018910032 0.012640532 0.012785627 0.008362832 0.012796397 0.0126071994
#> [22,] 0.018805612 0.011514207 0.005033262 0.009970615 0.012598593 0.0077731492
#> [23,] 0.017936785 0.016160621 0.009154123 0.011075093 0.012286358 0.0075551893
#> [24,] 0.009686801 0.002701354 0.022544394 0.015996773 0.009089200 0.0181390166
#> [25,] 0.024275376 0.011102967 0.002043382 0.002003565 0.001983833 0.0014173447
#>               V56           V57          V58         V59         V6         V60
#>  [1,] 0.006519486  0.0065530885 0.0055952867 0.006302200 0.10533770 0.006895639
#>  [2,] 0.003840712  0.0035912785 0.0038091370 0.003715991 0.09481376 0.004803648
#>  [3,] 0.007496761  0.0043255946 0.0128446083 0.014672061 0.20873391 0.008389206
#>  [4,] 0.015779743  0.0128214347 0.0144890631 0.015221358 0.14979991 0.010640411
#>  [5,] 0.013273688  0.0091951807 0.0105697189 0.005647331 0.11272590 0.008126550
#>  [6,] 0.007694558  0.0077508610 0.0026196135 0.002307778 0.07780623 0.001173937
#>  [7,] 0.003400110  0.0143740851 0.0154863359 0.018250409 0.15593817 0.019648464
#>  [8,] 0.006113923  0.0045552358 0.0052115800 0.007106656 0.09885552 0.007275971
#>  [9,] 0.009483037  0.0069150685 0.0157035720 0.014212006 0.06041277 0.006674617
#> [10,] 0.006734407  0.0058921091 0.0082642816 0.006192213 0.10376072 0.002999393
#> [11,] 0.017881788  0.0201141884 0.0066417279 0.013695271 0.10806939 0.013247191
#> [12,] 0.006540273  0.0062302730 0.0064437116 0.004957167 0.11209641 0.004964755
#> [13,] 0.005002116  0.0039747178 0.0053187276 0.006209816 0.15466887 0.005427972
#> [14,] 0.002441610  0.0043674351 0.0032226069 0.002285405 0.07172122 0.002640338
#> [15,] 0.005138779  0.0040209394 0.0031634286 0.003721786 0.10346070 0.004310255
#> [16,] 0.008627448  0.0072425768 0.0007633468 0.004963414 0.08451464 0.003272623
#> [17,] 0.007155854  0.0128036635 0.0172673024 0.011576158 0.05789451 0.005769048
#> [18,] 0.006162596  0.0038863076 0.0050531843 0.004750438 0.05491029 0.003202994
#> [19,] 0.005723564 -0.0007706701 0.0018323980 0.002313115 0.08964947 0.002387979
#> [20,] 0.004388814  0.0034109667 0.0004681026 0.002403258 0.03347991 0.002582962
#> [21,] 0.006466193  0.0114442757 0.0074861319 0.008096007 0.08639908 0.005356684
#> [22,] 0.005924635  0.0075833837 0.0079371791 0.009919930 0.07415353 0.005161570
#> [23,] 0.009554972  0.0091699164 0.0056477531 0.006583011 0.10688648 0.005538850
#> [24,] 0.024564862  0.0324740380 0.0165077588 0.009465312 0.23380104 0.007500045
#> [25,] 0.004749454  0.0132001554 0.0138874113 0.002611257 0.06237155 0.006439466
#>               V7         V8         V9
#>  [1,] 0.09963040 0.12910061 0.24185285
#>  [2,] 0.10920227 0.13048962 0.17701991
#>  [3,] 0.17277468 0.13658126 0.14981928
#>  [4,] 0.16946002 0.19853235 0.16043168
#>  [5,] 0.18668554 0.34644468 0.55436178
#>  [6,] 0.11137256 0.11044924 0.27596503
#>  [7,] 0.16859987 0.07378571 0.17032555
#>  [8,] 0.11847866 0.12858903 0.15792587
#>  [9,] 0.16289925 0.30369864 0.48787164
#> [10,] 0.13261628 0.15426235 0.17821258
#> [11,] 0.08615354 0.08258729 0.19699450
#> [12,] 0.10765854 0.08989169 0.12541774
#> [13,] 0.13230627 0.17949847 0.29214806
#> [14,] 0.07537209 0.07484934 0.10802579
#> [15,] 0.10765990 0.09854291 0.13311418
#> [16,] 0.12256806 0.09307496 0.11763879
#> [17,] 0.05397078 0.01847600 0.04740070
#> [18,] 0.08425765 0.05332378 0.04856789
#> [19,] 0.09742166 0.04115893 0.03821547
#> [20,] 0.08686975 0.07847202 0.08202815
#> [21,] 0.13557598 0.16659617 0.22107026
#> [22,] 0.08264942 0.06888078 0.05546036
#> [23,] 0.13511791 0.13631036 0.11772189
#> [24,] 0.24787810 0.37839760 0.55684360
#> [25,] 0.05810767 0.01605557 0.07748525
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986