Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.047543356 0.21930189 0.32597025 0.44542084 0.52282418 0.49714347
#>  [2,] 0.051675269 0.22770933 0.25444460 0.30655281 0.32075974 0.29227961
#>  [3,] 0.018554761 0.31389205 0.48189803 0.52747896 0.44218578 0.34339990
#>  [4,] 0.056722598 0.36856053 0.53561896 0.53715871 0.55502113 0.59040673
#>  [5,] 0.064141480 0.39587286 0.47775215 0.57727937 0.56635496 0.43729726
#>  [6,] 0.018470506 0.15073833 0.24205299 0.35742355 0.36596953 0.50200750
#>  [7,] 0.011891449 0.29197391 0.32054486 0.28439900 0.40032054 0.52439608
#>  [8,] 0.022266259 0.20703125 0.19803942 0.15729665 0.13211785 0.07717697
#>  [9,] 0.091788350 0.55597309 0.39632382 0.12619748 0.11817250 0.26071622
#> [10,] 0.042903754 0.71809151 0.74387158 0.50805410 0.30156455 0.19630867
#> [11,] 0.027735302 0.26355482 0.33745106 0.35199755 0.29137503 0.20356348
#> [12,] 0.024655895 0.38008765 0.46475575 0.51931157 0.51551598 0.41353442
#> [13,] 0.048155420 0.12092003 0.20543831 0.29138989 0.30589624 0.36968824
#> [14,] 0.039890863 0.23173074 0.29692545 0.34974097 0.32917499 0.30138556
#> [15,] 0.021551103 0.12338958 0.12615313 0.07598343 0.06997103 0.12553725
#> [16,] 0.010547456 0.08739875 0.07841044 0.07819420 0.06283001 0.11581805
#> [17,] 0.012232850 0.11266607 0.17052285 0.10679706 0.05972040 0.17838861
#> [18,] 0.009529589 0.09062505 0.11157769 0.07761423 0.15188658 0.21322378
#> [19,] 0.024933715 0.07946021 0.08204415 0.22069438 0.16944378 0.10963828
#> [20,] 0.015157500 0.05553042 0.18253189 0.39056371 0.29051457 0.25358396
#> [21,] 0.026702606 0.30630111 0.28513082 0.16218232 0.26092148 0.49185240
#> [22,] 0.029100000 0.11460000 0.09420000 0.07940000 0.02520000 0.11910000
#> [23,] 0.019514426 0.13382959 0.12962995 0.14148739 0.18808297 0.17695696
#> [24,] 0.020117833 0.09951351 0.08910591 0.08717680 0.11381467 0.05023086
#> [25,] 0.011289693 0.10709613 0.08737285 0.03534853 0.02077048 0.09400155
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.58697461 0.72034473 0.89062617 0.9294925 0.9127904 0.095350537
#>  [2,] 0.27515755 0.32271990 0.32532273 0.2135370 0.2879622 0.075150021
#>  [3,] 0.28231700 0.38809045 0.44499715 0.3936766 0.4044979 0.016676632
#>  [4,] 0.67987396 0.84590869 0.96259676 0.9310712 0.8219510 0.084279532
#>  [5,] 0.29096579 0.34431148 0.40500974 0.4022983 0.5202801 0.071790329
#>  [6,] 0.65306586 0.69562966 0.77445459 0.8718584 0.9580453 0.025234703
#>  [7,] 0.51808024 0.58791211 0.62851102 0.7318218 0.8503183 0.024830296
#>  [8,] 0.25457118 0.42853378 0.52195666 0.4952821 0.5707487 0.031135659
#>  [9,] 0.39982731 0.38187493 0.28205400 0.2553143 0.4078533 0.099194703
#> [10,] 0.27935236 0.37662542 0.24629731 0.2429623 0.3214735 0.091079642
#> [11,] 0.17588284 0.17763285 0.20031608 0.2361623 0.2947045 0.037738368
#> [12,] 0.29443918 0.26195114 0.29172812 0.3158530 0.4801634 0.022871336
#> [13,] 0.46619398 0.53590317 0.68976891 0.7943597 0.8591944 0.049304803
#> [14,] 0.28474037 0.31847840 0.36660710 0.5073553 0.6014757 0.040005305
#> [15,] 0.07508377 0.10161641 0.22653516 0.2397241 0.2765236 0.009753534
#> [16,] 0.04620067 0.13390169 0.14929310 0.3806527 0.3537555 0.007956246
#> [17,] 0.30404120 0.21987238 0.13896665 0.1710979 0.1760624 0.019697607
#> [18,] 0.21691364 0.20256184 0.15437383 0.1585958 0.1548760 0.008462302
#> [19,] 0.13680897 0.20703657 0.22355743 0.1810792 0.1928156 0.028167818
#> [20,] 0.12504074 0.08821192 0.16175569 0.4148678 0.2702460 0.060478124
#> [21,] 0.63249382 0.76745610 0.89611146 0.8335253 0.6027275 0.031743526
#> [22,] 0.10450000 0.20500000 0.15560000 0.2690000 0.3784000 0.040000000
#> [23,] 0.18434768 0.28150961 0.39376354 0.4425419 0.4835853 0.029774876
#> [24,] 0.01130101 0.02238316 0.06671406 0.2035151 0.3644230 0.003822107
#> [25,] 0.08284395 0.09148387 0.20696423 0.2965015 0.3783272 0.002832948
#>             V20       V21       V22        V23        V24        V25       V26
#>  [1,] 0.8330634 0.7682663 0.5779660 0.40940467 0.39547837 0.26772950 0.1359194
#>  [2,] 0.3759185 0.4037516 0.4288302 0.49204393 0.53424450 0.55173730 0.6181357
#>  [3,] 0.4932925 0.7261674 0.9452041 1.00530445 0.98419214 0.99705491 0.9176642
#>  [4,] 0.7653746 0.6970012 0.4237049 0.12184991 0.07301679 0.26289102 0.2974586
#>  [5,] 0.7076119 0.7694811 0.8739513 0.87919086 0.92394286 1.01320010 0.9741743
#>  [6,] 1.0155676 0.9106207 0.8906747 0.93587713 0.85228357 0.57050737 0.4124126
#>  [7,] 0.9144636 0.9288908 0.9844368 1.00833012 0.87266512 0.71780570 0.7296203
#>  [8,] 0.7677293 0.9267417 1.0133421 0.89198301 0.71759529 0.51181540 0.2639680
#>  [9,] 0.8633292 0.9807388 0.4270012 0.08934259 0.17490975 0.14556082 0.3155413
#> [10,] 0.4801658 0.6486890 0.8468347 0.97639120 1.00588890 0.67219193 0.5384038
#> [11,] 0.2983515 0.4152587 0.5467555 0.63163159 0.70163464 0.72448717 0.7655507
#> [12,] 0.6815282 0.7730881 0.8521090 0.79225435 0.80082755 0.83915269 0.9053903
#> [13,] 0.8670947 0.9056981 0.9084977 0.88960777 0.90367623 0.80570987 0.6534926
#> [14,] 0.6927870 0.7815404 0.7979065 0.82070233 0.87298482 0.86124333 0.9688078
#> [15,] 0.4072253 0.5474385 0.7157686 0.87689439 0.89655002 0.86909806 0.9290401
#> [16,] 0.3162322 0.2116985 0.2219312 0.24523688 0.25236618 0.06938896 0.2342191
#> [17,] 0.2406605 0.2418841 0.4734692 0.65794682 0.64287493 0.69407138 0.7887679
#> [18,] 0.1511589 0.1822064 0.3387006 0.49310038 0.62960591 0.68014200 0.7683627
#> [19,] 0.2263411 0.2384215 0.1099403 0.17385631 0.14738287 0.08195198 0.2119527
#> [20,] 0.6138733 0.7740804 0.7616277 0.75286287 0.80497044 1.00002846 0.9434046
#> [21,] 0.4886418 0.5741912 0.7198438 0.78136949 0.82045167 0.74391656 0.6875833
#> [22,] 0.4024000 0.3470000 0.1395000 0.12080000 0.28270000 0.15000000 0.2626000
#> [23,] 0.5174923 0.5679363 0.5508702 0.57808274 0.62351863 0.64724398 0.6606044
#> [24,] 0.3951495 0.3696738 0.4448411 0.59612716 0.72589029 0.80846614 0.8057024
#> [25,] 0.5492949 0.6501383 0.7757200 0.86679395 0.80963450 0.86411267 0.9737558
#>              V27        V28        V29           V3       V30       V31
#>  [1,] 0.12843165 0.13776695 0.16679683  0.101903436 0.1708475 0.2766843
#>  [2,] 0.75698051 0.88170322 0.84027312  0.061749380 0.8272619 0.8225642
#>  [3,] 0.82264443 0.69713956 0.45620817  0.028184064 0.3317205 0.2542014
#>  [4,] 0.27271624 0.23934722 0.09127697  0.125572521 0.2160784 0.1950193
#>  [5,] 0.98745350 0.74301757 0.42238623  0.043148831 0.3018369 0.3388499
#>  [6,] 0.33962338 0.38018679 0.42206335  0.037776911 0.3989628 0.2682140
#>  [7,] 0.74045281 0.73537845 0.67764765  0.041388436 0.5119540 0.2971807
#>  [8,] 0.07452903 0.00748478 0.21018201  0.013706409 0.3842197 0.3279849
#>  [9,] 0.19452939 0.36414032 0.49979553  0.081283813 0.6968985 0.2984804
#> [10,] 0.40381459 0.34363003 0.42334267  0.083331336 0.4135019 0.4324598
#> [11,] 0.87485643 0.94453840 0.87414144  0.055578134 0.7245435 0.6866483
#> [12,] 0.97003272 0.87110076 0.67012641  0.031364386 0.5043398 0.4518474
#> [13,] 0.52498873 0.36566181 0.33680508  0.076102181 0.3142829 0.2529430
#> [14,] 0.98953794 0.89314353 0.78418368  0.048934449 0.6433380 0.4420921
#> [15,] 0.94629235 0.89324093 0.88370895  0.014993877 0.7675004 0.6820015
#> [16,] 0.43301951 0.56417034 0.67529565 -0.001934737 0.6487591 0.5257221
#> [17,] 0.72497123 0.75119924 0.74658525  0.019585286 0.6790742 0.5025427
#> [18,] 0.80104697 0.89018430 0.89989675  0.024672266 0.8128605 0.7018825
#> [19,] 0.43977633 0.63513897 0.67797019  0.041442882 0.6644551 0.7336397
#> [20,] 0.73536488 0.57586759 0.55787332  0.070033090 0.4793340 0.5343436
#> [21,] 0.68546554 0.58200777 0.47487997  0.039396345 0.5208182 0.4685378
#> [22,] 0.44680000 0.75200000 0.90360000  0.077100000 0.7812000 0.4766000
#> [23,] 0.66171288 0.64262022 0.62416534  0.032238551 0.6341629 0.6150884
#> [24,] 0.70613447 0.79423496 0.97395322  0.013811032 0.9735900 0.9286537
#> [25,] 0.99164173 0.90599186 0.78699288  0.005185430 0.6770315 0.5892720
#>             V32        V33        V34         V35       V36       V37       V38
#>  [1,] 0.3321710 0.34423122 0.45680793  0.49275777 0.2997082 0.2139321 0.3095085
#>  [2,] 0.7582159 0.52718019 0.47694013  0.47580174 0.3688355 0.3499598 0.3858517
#>  [3,] 0.4019521 0.42065956 0.32963970  0.14080800 0.1937512 0.3229494 0.3357641
#>  [4,] 0.1965072 0.25160893 0.20283056  0.27551267 0.1123289 0.1289524 0.1972622
#>  [5,] 0.3859550 0.35860379 0.27997921  0.07090495 0.3871872 0.4800190 0.3463557
#>  [6,] 0.2629256 0.28978367 0.22309608  0.20689653 0.2168407 0.2567664 0.4090708
#>  [7,] 0.1927924 0.05609938 0.09887326  0.25967288 0.3643918 0.2792979 0.2349324
#>  [8,] 0.5575350 0.44169397 0.15028857 -0.07068949 0.1898614 0.4185976 0.4000823
#>  [9,] 0.3299414 0.55734482 0.72113644  0.85021342 0.8422488 0.7775941 0.8927484
#> [10,] 0.1023988 0.42606149 0.65235219  0.61747755 0.4339717 0.4067258 0.3559572
#> [11,] 0.5823317 0.44589976 0.33386357  0.18752419 0.1001325 0.2266445 0.1454475
#> [12,] 0.5110187 0.46473327 0.32981432  0.10984317 0.2916824 0.3453270 0.3349819
#> [13,] 0.2049857 0.30514714 0.41194911  0.49739419 0.4549253 0.3398918 0.3347696
#> [14,] 0.3159410 0.24087379 0.10292923  0.05531991 0.0373356 0.1023789 0.1486946
#> [15,] 0.6373720 0.61082421 0.65490962  0.66451863 0.5680566 0.4686006 0.3724652
#> [16,] 0.2566097 0.49469159 0.44367035  0.25024477 0.3341471 0.4834985 0.6349095
#> [17,] 0.3352085 0.55623537 0.87958522  0.85069014 0.6826947 0.3729967 0.3324742
#> [18,] 0.5963999 0.56955298 0.54833207  0.49066917 0.4062873 0.2125149 0.1448743
#> [19,] 0.6124668 0.63469850 0.84134723  0.94058996 0.9401715 0.7926696 0.6935570
#> [20,] 0.5339211 0.66244645 0.56700947  0.59143693 0.6232512 0.5623225 0.2526476
#> [21,] 0.3278287 0.27474973 0.19459104  0.41203001 0.5542333 0.5347238 0.3067577
#> [22,] 0.2483000 0.53720000 0.62790000  0.36470000 0.4572000 0.6359000 0.6474000
#> [23,] 0.5872979 0.59276106 0.62830689  0.63245960 0.6177629 0.5557055 0.5202130
#> [24,] 0.8870801 0.93547441 0.91058257  0.73736163 0.4995329 0.2870015 0.1664079
#> [25,] 0.5462531 0.55005889 0.56277185  0.55439133 0.5067163 0.4568110 0.3555642
#>                V39            V4        V40       V41        V42        V43
#>  [1,]  0.307629650  1.189263e-01 0.20058062 0.1562055 0.11069528 0.10797351
#>  [2,]  0.356233887  5.209599e-02 0.29619982 0.5341833 0.62053724 0.53609010
#>  [3,]  0.570618929  3.007008e-02 0.55541879 0.3265576 0.04189427 0.31518122
#>  [4,]  0.175427264  1.238600e-01 0.04572368 0.1088580 0.17406280 0.16259593
#>  [5,]  0.168035379  3.337742e-02 0.10643589 0.1223816 0.26445025 0.46037845
#>  [6,]  0.309581238  9.883980e-02 0.21683774 0.1631161 0.14439979 0.19440434
#>  [7,]  0.278133992  7.207494e-02 0.28064148 0.3667425 0.40560391 0.31640805
#>  [8,]  0.292226113  4.848809e-02 0.17063558 0.1661023 0.08187485 0.01122076
#>  [9,]  0.826976929  7.319694e-02 0.58180902 0.3502055 0.20700986 0.18492942
#> [10,]  0.453363597  8.031740e-02 0.43457588 0.2939267 0.43684269 0.42116283
#> [11,]  0.149230956  6.903692e-02 0.24316750 0.3303721 0.31961860 0.25029834
#> [12,]  0.426052529  2.917206e-02 0.38395837 0.1423008 0.27023070 0.36522941
#> [13,]  0.390391719  1.000623e-01 0.36338675 0.3656192 0.38567467 0.31748888
#> [14,]  0.163866211  6.760366e-02 0.20195520 0.1740936 0.20658606 0.22803568
#> [15,]  0.238342869  2.359431e-02 0.21349329 0.1929723 0.09842579 0.10690059
#> [16,]  0.725960810 -8.997897e-05 0.59604603 0.2763837 0.25876602 0.19277839
#> [17,]  0.242796927  4.839407e-02 0.30642041 0.3358140 0.39300191 0.42512885
#> [18,]  0.175501547  2.706285e-02 0.15687750 0.1447112 0.15862888 0.22988669
#> [19,]  0.645284082  2.640794e-02 0.64557255 0.5628535 0.49755050 0.35569196
#> [20,] -0.009722106  6.137599e-02 0.32245269 0.4118956 0.60606878 0.50598163
#> [21,]  0.160852491  1.476629e-02 0.30411406 0.3126346 0.29713091 0.20886142
#> [22,]  0.552000000  8.090000e-02 0.32530000 0.2292000 0.06530000 0.00000000
#> [23,]  0.471894749  2.892116e-02 0.37529141 0.2338512 0.16526619 0.16753739
#> [24,]  0.089890300  1.250444e-02 0.25022857 0.3155202 0.34839884 0.30555291
#> [25,]  0.150082754  4.186909e-02 0.14915402 0.1657590 0.06535050 0.01344571
#>                V44        V45          V46          V47        V48         V49
#>  [1,]  0.073435015 0.04507499  0.051822157  0.088046206 0.08620253 0.059848514
#>  [2,]  0.444110023 0.54466827  0.530542598  0.361935262 0.23252616 0.133099016
#>  [3,]  0.341941514 0.41629803  0.366026101  0.208469225 0.08690570 0.060523629
#>  [4,]  0.130661114 0.08778356  0.033929067  0.068940916 0.07245998 0.021073667
#>  [5,]  0.490788158 0.33247701  0.137118393  0.034935453 0.04144056 0.031243780
#>  [6,]  0.163110940 0.09142180 -0.008698674 -0.009310611 0.04196730 0.016828294
#>  [7,]  0.241500495 0.15280663  0.142037973  0.127609925 0.12119930 0.074177056
#>  [8,]  0.064376561 0.04011360  0.026457144  0.022853704 0.02754997 0.005024297
#>  [9,]  0.268997790 0.47178422  0.377932625  0.194783605 0.19411807 0.103931275
#> [10,]  0.180826493 0.14776888  0.010431397  0.117653434 0.20932514 0.108316828
#> [11,]  0.137991960 0.13801629  0.189335765  0.197707268 0.17128061 0.083978407
#> [12,]  0.450890597 0.44485184  0.224322475  0.117756159 0.10126963 0.047926069
#> [13,]  0.287795836 0.22329888  0.086354514  0.087892547 0.05152180 0.029868938
#> [14,]  0.202272771 0.15936421  0.156082420  0.138246818 0.08842778 0.049126544
#> [15,]  0.075715628 0.03560968  0.034093270  0.028026189 0.03693745 0.010031157
#> [16,]  0.097394670 0.06331036  0.037076567  0.090870659 0.08386250 0.023860449
#> [17,]  0.235356946 0.09781416  0.063084956  0.100462696 0.07195219 0.033609113
#> [18,]  0.174812011 0.07265067  0.053610065  0.035957449 0.03652499 0.022717818
#> [19,]  0.215858143 0.12996108  0.071295378  0.095054410 0.09760321 0.051589716
#> [20,]  0.301107454 0.15908715  0.097991981  0.061392883 0.06868847 0.016576143
#> [21,]  0.164654199 0.17197921  0.137552610  0.130560673 0.11898429 0.074748676
#> [22,]  0.000000000 0.00000000  0.000000000  0.000000000 0.00000000 0.000000000
#> [23,]  0.156931759 0.10141705  0.089153151  0.053076175 0.02958839 0.020078809
#> [24,]  0.300594868 0.28343635  0.199281319  0.092385057 0.02795548 0.023919951
#> [25,] -0.003217946 0.04838961  0.006725554  0.009593499 0.01877189 0.003738777
#>               V5         V50         V51          V52          V53          V54
#>  [1,] 0.12020266 0.022389035 0.018250320  0.028791976  0.011900901 0.0124225807
#>  [2,] 0.07805888 0.049645317 0.049529770  0.033324866  0.015599447 0.0168337756
#>  [3,] 0.04514469 0.026847946 0.012621349  0.003931845  0.013626597 0.0138529330
#>  [4,] 0.14121573 0.014042874 0.027078917  0.014453116  0.026030496 0.0113370706
#>  [5,] 0.01311931 0.008091240 0.015964060  0.004617510  0.001974650 0.0089322598
#>  [6,] 0.16690676 0.004705465 0.007921710  0.018949711  0.010842871 0.0069685877
#>  [7,] 0.15249405 0.019677882 0.022133635  0.021667036  0.018285562 0.0252351922
#>  [8,] 0.04627615 0.003123026 0.011502617  0.006476510 -0.003460964 0.0030744617
#>  [9,] 0.04626039 0.045205067 0.026094516  0.015707358  0.014255199 0.0165590542
#> [10,] 0.01287847 0.019953719 0.021026066  0.017561949  0.019876824 0.0213267881
#> [11,] 0.08252290 0.016860728 0.022587376  0.017733962  0.008609767 0.0082868819
#> [12,] 0.06418781 0.009697497 0.016050849  0.013000513  0.007950522 0.0095924593
#> [13,] 0.10369523 0.007480500 0.012031089  0.004454343  0.013881367 0.0146160176
#> [14,] 0.09010242 0.015462875 0.014003000  0.011602415  0.007077405 0.0079290182
#> [15,] 0.05515355 0.008019005 0.004756836  0.004309058  0.009658763 0.0079597260
#> [16,] 0.02859074 0.014199041 0.008943006  0.020161501  0.024708846 0.0164157237
#> [17,] 0.06463498 0.020081724 0.031971802  0.014803901  0.011861329 0.0108664007
#> [18,] 0.02273239 0.010264503 0.013951579  0.012307813  0.013364804 0.0038987772
#> [19,] 0.02593501 0.020161374 0.012132002  0.005666760  0.015965781 0.0130532919
#> [20,] 0.13778615 0.004502422 0.002637544 -0.001242706  0.012008021 0.0105152980
#> [21,] 0.03810553 0.017546840 0.017384781  0.008419551  0.006032602 0.0109117641
#> [22,] 0.05210000 0.000000000 0.000000000  0.005600000  0.023700000 0.0204000000
#> [23,] 0.04701508 0.022687649 0.012854066  0.010844685  0.007710810 0.0138157894
#> [24,] 0.01458766 0.011879615 0.007450911  0.008613670  0.006958946 0.0073727125
#> [25,] 0.05025428 0.010414816 0.008118431  0.008240340  0.007151761 0.0008216592
#>                V55          V56          V57         V58          V59
#>  [1,] 1.865020e-02 0.0117785323 0.0038742732 0.010582042 0.0184268616
#>  [2,] 1.313065e-02 0.0102510690 0.0099128696 0.011311156 0.0089077655
#>  [3,] 1.728478e-02 0.0132581810 0.0110415579 0.012324674 0.0115097128
#>  [4,] 1.542497e-02 0.0173860542 0.0177159629 0.012264384 0.0123746283
#>  [5,] 6.929030e-03 0.0094889400 0.0086055300 0.008710140 0.0039557600
#>  [6,] 2.490375e-02 0.0241701303 0.0068052499 0.005047108 0.0149756319
#>  [7,] 1.277690e-02 0.0070879285 0.0074156288 0.021496281 0.0100207081
#>  [8,] 2.174404e-03 0.0027683035 0.0082907511 0.003878694 0.0007371139
#>  [9,] 1.664187e-02 0.0169612865 0.0182625060 0.013260209 0.0087721077
#> [10,] 1.432159e-02 0.0048888086 0.0026507580 0.016379062 0.0059693865
#> [11,] 5.483994e-03 0.0067152140 0.0059564924 0.007666439 0.0054009685
#> [12,] 6.908032e-03 0.0045931008 0.0040886580 0.003509842 0.0073133541
#> [13,] 6.501767e-03 0.0153238905 0.0132240200 0.009663671 0.0187527018
#> [14,] 3.323380e-03 0.0046774606 0.0041589172 0.002473371 0.0040438209
#> [15,] 5.056167e-03 0.0064414831 0.0063148409 0.005303424 0.0025517555
#> [16,] 2.313130e-02 0.0034297038 0.0008697282 0.004509308 0.0078364235
#> [17,] 3.780357e-03 0.0089136870 0.0095586901 0.003678376 0.0065976121
#> [18,] 4.696972e-03 0.0043508827 0.0085259191 0.004545728 0.0076242572
#> [19,] 7.763967e-03 0.0063537299 0.0076747435 0.013111533 0.0179730980
#> [20,] 1.167446e-02 0.0109646591 0.0118036376 0.007476063 0.0035088867
#> [21,] 8.088865e-03 0.0059807709 0.0111021339 0.008662187 0.0107848304
#> [22,] 5.000000e-03 0.0137000000 0.0164000000 0.008100000 0.0139000000
#> [23,] 1.203080e-02 0.0085839252 0.0112861343 0.008807592 0.0044891760
#> [24,] 2.164400e-03 0.0007560619 0.0013998607 0.002800112 0.0060660454
#> [25,] 6.359806e-05 0.0058203981 0.0028875585 0.001339751 0.0018901823
#>               V6           V60         V7         V8          V9
#>  [1,] 0.05326489  0.0096079125 0.08413472 0.07742615 0.121095257
#>  [2,] 0.11748446  0.0067235505 0.12144753 0.16677605 0.229862639
#>  [3,] 0.06861434  0.0068249347 0.05685030 0.11111453 0.243892498
#>  [4,] 0.14000557  0.0134204523 0.12228261 0.10942489 0.222600173
#>  [5,] 0.12786914  0.0050889400 0.10143364 0.19901665 0.346690403
#>  [6,] 0.11823000  0.0091336986 0.11216576 0.13624658 0.103687340
#>  [7,] 0.18876521  0.0026977869 0.16877651 0.17685355 0.224354124
#>  [8,] 0.08329176 -0.0003160554 0.16883932 0.15509762 0.221458721
#>  [9,] 0.10674819  0.0099427975 0.21236354 0.29623245 0.494144018
#> [10,] 0.02144850  0.0020865704 0.05144085 0.20501733 0.388827582
#> [11,] 0.12530912  0.0041641822 0.11897211 0.11658083 0.183609830
#> [12,] 0.11151917  0.0043930993 0.12749154 0.16907547 0.305216367
#> [13,] 0.13968861  0.0136280118 0.15584482 0.13789155 0.086230027
#> [14,] 0.10735890  0.0039598518 0.15189383 0.16076777 0.193936997
#> [15,] 0.09023448  0.0032331058 0.08304847 0.08441543 0.133054597
#> [16,] 0.04351469  0.0093496257 0.04261772 0.08017245 0.126018805
#> [17,] 0.08312965  0.0095041260 0.10524347 0.09753483 0.084213135
#> [18,] 0.05101668  0.0068779631 0.02548391 0.05591672 0.094672901
#> [19,] 0.04730577  0.0061836467 0.04307454 0.07146073 0.067992534
#> [20,] 0.20141290  0.0156363937 0.14688417 0.13623868 0.005494727
#> [21,] 0.09805659  0.0075865124 0.16204993 0.22775702 0.261434428
#> [22,] 0.10510000  0.0111000000 0.01450000 0.06740000 0.129400000
#> [23,] 0.09549333  0.0028849284 0.13017607 0.10369758 0.133461756
#> [24,] 0.01208381  0.0036768959 0.05687635 0.03011031 0.045198792
#> [25,] 0.07614272  0.0022379420 0.08856955 0.08609409 0.154605770
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406