Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1        V10         V11          V12        V13        V14
#>  [1,]  0.085019039 0.17385930  0.29550898  0.407664123 0.54767663 0.68130996
#>  [2,]  0.019634372 0.40274728  0.38753135  0.452862400 0.46230546 0.50218958
#>  [3,]  0.023541968 0.27385086  0.32320696  0.338269852 0.27273524 0.15722986
#>  [4,]  0.052825340 0.32474743  0.29038605  0.239197473 0.25912901 0.26653872
#>  [5,]  0.022177735 0.21824493  0.20760043  0.155124050 0.13495784 0.10242800
#>  [6,]  0.019891705 0.16952876  0.32165779  0.482991384 0.54724915 0.54396013
#>  [7,]  0.021297795 0.20551427  0.25182495  0.294150211 0.29933239 0.26225692
#>  [8,]  0.059172899 0.28925178  0.39056783  0.447222679 0.51950523 0.49043503
#>  [9,]  0.005211618 0.20185600  0.06604251 -0.006217002 0.21823556 0.33230633
#> [10,]  0.021588971 0.06386046  0.10889313  0.156094678 0.22698275 0.27618447
#> [11,]  0.032820118 0.24752319  0.31899604  0.380884072 0.38930864 0.44035306
#> [12,]  0.010893649 0.13669439  0.21263945  0.241049257 0.22480039 0.14291364
#> [13,]  0.021253888 0.42600120  0.50550694  0.505406121 0.49079679 0.37584177
#> [14,]  0.053395967 0.10693907  0.24507766  0.335244252 0.32551110 0.37979407
#> [15,]  0.013984726 0.14147723  0.14368800  0.096010965 0.06911741 0.11923619
#> [16,]  0.015217582 0.06870448  0.11992290  0.164099575 0.27447987 0.30162035
#> [17,]  0.012378110 0.11255256  0.10388651  0.097035411 0.08581470 0.13156489
#> [18,]  0.019691646 0.08143348  0.07015372  0.073916185 0.10990496 0.05298405
#> [19,]  0.044276448 0.07049919  0.13297522  0.247944937 0.25083664 0.12529773
#> [20,]  0.009502767 0.10550504  0.09954584  0.093294545 0.10537022 0.11667220
#> [21,]  0.027512892 0.04492707  0.02482019  0.177318933 0.13893506 0.09032428
#> [22,] -0.012384037 0.02389827 -0.01509384  0.169507789 0.18653430 0.08855284
#> [23,]  0.017770921 0.25009243  0.23191719  0.223197454 0.31453703 0.41998989
#> [24,]  0.012106353 0.15561388  0.19132487  0.138465590 0.18779651 0.11401138
#> [25,]  0.015029969 0.10309473  0.11592729  0.094911816 0.12681624 0.18922663
#>               V15        V16        V17       V18        V19            V2
#>  [1,] 0.866124206 0.85626777 0.80629544 0.8962417 1.00880859  0.0996237315
#>  [2,] 0.480339582 0.60906486 0.83246214 0.8543340 0.84469583  0.0991941496
#>  [3,] 0.157077788 0.20530006 0.23023191 0.3302276 0.45395221  0.0339077641
#>  [4,] 0.289228532 0.32189206 0.27226457 0.1905222 0.25787056  0.0651918128
#>  [5,] 0.286242694 0.46107648 0.55980419 0.5049955 0.57561520  0.0313825449
#>  [6,] 0.612119368 0.73924290 0.87779533 0.9795243 0.98210136  0.0322411051
#>  [7,] 0.203039911 0.16765532 0.17704445 0.2281266 0.32239497  0.0383865853
#>  [8,] 0.593427260 0.71350529 0.90452969 0.9242260 0.88321375  0.1101979364
#>  [9,] 0.451109808 0.61493268 0.63814072 0.6962297 0.82328932 -0.0013280311
#> [10,] 0.328440405 0.48222095 0.58229669 0.6473584 0.81554808  0.0053981390
#> [11,] 0.470344879 0.45984444 0.52305331 0.6246556 0.71183218  0.0302226550
#> [12,] 0.060204662 0.15180914 0.30199617 0.4034804 0.52273276  0.0143956913
#> [13,] 0.249872733 0.23556329 0.27191925 0.3121014 0.49742442  0.0238518069
#> [14,] 0.530074844 0.54653376 0.74373576 0.8860238 0.98283182  0.0493489146
#> [15,] 0.103593563 0.10502019 0.22234243 0.2849906 0.35637411 -0.0001387093
#> [16,] 0.284204194 0.31132069 0.23840071 0.2919277 0.48433782  0.0190189631
#> [17,] 0.074080275 0.07618707 0.20261911 0.3198439 0.31923605  0.0231068846
#> [18,] 0.008541407 0.02515939 0.07712358 0.2166066 0.36504598  0.0047529368
#> [19,] 0.130891664 0.22620723 0.24462201 0.3607591 0.23081588  0.0436247792
#> [20,] 0.172453636 0.17769163 0.14360214 0.1500648 0.10574954  0.0137648007
#> [21,] 0.108146626 0.18671243 0.18103965 0.1422019 0.16098453  0.0362816114
#> [22,] 0.107317468 0.09252965 0.08355513 0.2143291 0.05904386 -0.0011684690
#> [23,] 0.515158934 0.60570662 0.53525525 0.3988252 0.37350170  0.0495905433
#> [24,] 0.237735485 0.48597588 0.69461743 0.7833760 0.67346555  0.0213646059
#> [25,] 0.168759200 0.18799092 0.22765537 0.2281886 0.23368201  0.0121527136
#>               V20       V21        V22        V23       V24        V25
#>  [1,] 0.864866261 0.8583833 0.92830866 1.00340158 0.8899829 0.62915620
#>  [2,] 0.872517216 0.7873870 0.84475226 1.01966420 1.0678130 0.98766908
#>  [3,] 0.545714533 0.6554133 0.73514710 0.80082870 0.8865869 0.93465847
#>  [4,] 0.448041119 0.4924386 0.38505377 0.36335467 0.4137656 0.45501240
#>  [5,] 0.775117895 0.9397503 1.01554126 0.87300732 0.6828713 0.48334183
#>  [6,] 0.935672543 0.7958065 0.69266773 0.64789761 0.5729972 0.35938963
#>  [7,] 0.355525030 0.4170486 0.51197217 0.56969267 0.6584127 0.72538441
#>  [8,] 0.793246351 0.7868439 0.51780977 0.20992539 0.2005523 0.22534644
#>  [9,] 0.886676284 1.0016613 1.00652252 0.91267899 0.6289527 0.41538791
#> [10,] 0.913354769 0.9123152 0.96014097 1.02642206 0.9402461 0.88867619
#> [11,] 0.748383353 0.7941633 0.88310617 0.91272678 0.9677908 0.94225370
#> [12,] 0.582794265 0.6773527 0.76098156 0.76010451 0.7447148 0.81664080
#> [13,] 0.693576938 0.8045626 0.83447967 0.77390650 0.7986151 0.80196775
#> [14,] 1.018455772 1.0426668 1.02833087 0.95813305 0.8983410 0.76224132
#> [15,] 0.531553969 0.6536259 0.78509853 0.92232386 0.9025715 0.91792276
#> [16,] 0.635730280 0.6374277 0.65402229 0.68914413 0.6989760 0.71233422
#> [17,] 0.479488839 0.6147009 0.70926991 0.80637424 0.8290857 0.91568434
#> [18,] 0.395623421 0.3771875 0.44929837 0.58047553 0.7035905 0.79009523
#> [19,] 0.372122019 0.6473889 0.70066010 0.78654743 0.8198157 0.77479856
#> [20,] 0.131328484 0.2127098 0.24998349 0.37843984 0.5959124 0.82637960
#> [21,] 0.138054188 0.2005959 0.07671371 0.19691271 0.2251522 0.05475541
#> [22,] 0.002741245 0.1403653 0.09581817 0.08564985 0.1367004 0.05650017
#> [23,] 0.406835260 0.5142739 0.61703713 0.65766989 0.6193089 0.65982104
#> [24,] 0.440916775 0.4438461 0.26440408 0.38101764 0.7157695 0.76967080
#> [25,] 0.291251087 0.3856659 0.50349229 0.61260235 0.6906140 0.72377200
#>              V26        V27        V28        V29          V3       V30
#>  [1,] 0.57824838 0.44182804 0.39064818 0.54370107 0.095535838 0.2898267
#>  [2,] 0.98877514 0.95678900 0.71943251 0.65271734 0.111946969 0.5691513
#>  [3,] 0.97989106 0.98183520 0.93478114 0.83258776 0.033145466 0.6866363
#>  [4,] 0.55574429 0.58161842 0.72715682 0.75050781 0.062893117 0.8229861
#>  [5,] 0.29528010 0.14211705 0.03593233 0.18988486 0.014750031 0.3875621
#>  [6,] 0.09856312 0.05134784 0.06520896 0.09799251 0.036846357 0.1893216
#>  [7,] 0.77471153 0.89699306 0.97495291 0.90360668 0.038492332 0.6771878
#>  [8,] 0.18135942 0.19434799 0.20387553 0.12585408 0.128113219 0.1463479
#>  [9,] 0.41857329 0.61506833 0.83267962 0.79680479 0.014944450 0.5710664
#> [10,] 0.83241543 0.79416151 0.75528086 0.61189261 0.018700339 0.4374630
#> [11,] 0.95282916 0.94092941 0.85949923 0.75646306 0.039610647 0.5726860
#> [12,] 0.93313572 0.98593885 0.96368573 0.84923657 0.035156241 0.6143182
#> [13,] 0.87301629 0.94045452 0.89653320 0.72071457 0.040373307 0.4856427
#> [14,] 0.63461005 0.52343267 0.37543099 0.35988815 0.085702027 0.2966954
#> [15,] 0.99914517 0.98560412 0.91139472 0.82540565 0.007110546 0.7046885
#> [16,] 0.76960711 0.82963725 0.81671791 0.69694822 0.022583176 0.5444079
#> [17,] 0.96970752 0.94981174 0.82013471 0.71792114 0.028507490 0.6426367
#> [18,] 0.79488652 0.71670629 0.80648839 0.97738637 0.014538725 0.9687439
#> [19,] 0.69452697 0.67769027 0.48873427 0.36886990 0.030503292 0.3044617
#> [20,] 0.97462266 0.83880570 0.61074753 0.49558580 0.011566546 0.3822551
#> [21,] 0.18319306 0.51897307 0.78067834 0.82124403 0.018759167 0.7255998
#> [22,] 0.15544515 0.32224398 0.53416570 0.57417952 0.003671154 0.5529734
#> [23,] 0.74085273 0.72516535 0.68950490 0.47818144 0.073895430 0.4950103
#> [24,] 0.73242144 0.74445512 0.79282348 0.57782095 0.036258350 0.3877734
#> [25,] 0.78105508 0.81187177 0.85828411 0.90729508 0.017303460 0.8574950
#>             V31        V32        V33        V34         V35        V36
#>  [1,] 0.2076060 0.07753693 0.02246355 0.12830006  0.25240988 0.16533578
#>  [2,] 0.2870247 0.21583328 0.11830422 0.34060494  0.40455129 0.24179868
#>  [3,] 0.5443752 0.39474235 0.27481351 0.18904128  0.15229498 0.07868027
#>  [4,] 0.7300488 0.68314005 0.60743604 0.59694140  0.63341741 0.56848999
#>  [5,] 0.3034341 0.53208846 0.44392511 0.15418304 -0.05833147 0.20333428
#>  [6,] 0.3058180 0.41490249 0.36759172 0.48244315  0.45868123 0.34325953
#>  [7,] 0.6172009 0.49928465 0.37933973 0.29065591  0.16602520 0.14385322
#>  [8,] 0.2035355 0.25340577 0.32707460 0.38483065  0.48805849 0.16939729
#>  [9,] 0.2579022 0.28255047 0.29444050 0.15282342  0.14268657 0.12211365
#> [10,] 0.2298845 0.13255427 0.15777861 0.25529415  0.37444171 0.38141345
#> [11,] 0.4147470 0.28615296 0.17205282 0.04526712  0.09950006 0.14285614
#> [12,] 0.2312139 0.22381226 0.33260393 0.51419192  0.75062226 0.87918321
#> [13,] 0.4031066 0.47844080 0.47336779 0.36752638  0.15887391 0.30354690
#> [14,] 0.2045937 0.13899039 0.24356092 0.39177875  0.41805394 0.36336751
#> [15,] 0.6020405 0.56203549 0.53586304 0.56201702  0.57543385 0.53985783
#> [16,] 0.4686249 0.34288831 0.28365601 0.24137352  0.22708437 0.30293965
#> [17,] 0.5808924 0.52482438 0.52049885 0.52889541  0.54945295 0.53857202
#> [18,] 0.9147503 0.87593514 0.93533232 0.90196377  0.73201132 0.49806466
#> [19,] 0.4019699 0.58853021 0.69664212 0.64071435  0.63612179 0.65111326
#> [20,] 0.2975926 0.16218651 0.18158021 0.21630920  0.26083121 0.19201549
#> [21,] 0.7360173 0.51649678 0.60941737 0.79954008  0.83182350 0.81483250
#> [22,] 0.6495250 0.65138255 0.71487280 0.88418536  0.95243577 0.93720810
#> [23,] 0.5819370 0.46014954 0.25648987 0.23661981  0.50052164 0.60417646
#> [24,] 0.4008907 0.44123302 0.35292454 0.16184972  0.26171448 0.34976845
#> [25,] 0.7440676 0.64042510 0.61779798 0.60345737  0.59687803 0.52307845
#>              V37        V38        V39          V4       V40        V41
#>  [1,] 0.14952440 0.08849995 0.34351260 0.069003422 0.2943333 0.08096308
#>  [2,] 0.19512901 0.26055989 0.23244420 0.036785846 0.2065179 0.36368610
#>  [3,] 0.10076866 0.10917195 0.12259824 0.040361506 0.1014328 0.08398526
#>  [4,] 0.49333787 0.53709206 0.47746118 0.068002048 0.4005902 0.47767768
#>  [5,] 0.43195437 0.40193426 0.28656803 0.044762159 0.1672449 0.15542684
#>  [6,] 0.29122184 0.41204320 0.31609126 0.042746073 0.1980965 0.21929139
#>  [7,] 0.22732749 0.14437846 0.19786571 0.044201643 0.2842621 0.30302144
#>  [8,] 0.10920224 0.24616786 0.26730989 0.160175542 0.1250938 0.06675322
#>  [9,] 0.06440759 0.07747127 0.09268761 0.066717207 0.1294319 0.23081934
#> [10,] 0.31660165 0.21877426 0.32610055 0.100288239 0.3994172 0.43233361
#> [11,] 0.16694949 0.25208000 0.28063928 0.063087960 0.2754148 0.28024926
#> [12,] 0.79380066 0.64222894 0.45763717 0.048511121 0.3220317 0.31596070
#> [13,] 0.31265470 0.32725279 0.49680579 0.040645029 0.4264503 0.18360687
#> [14,] 0.38312971 0.35007059 0.42144673 0.090324815 0.3226312 0.25725293
#> [15,] 0.48232057 0.37733140 0.17555720 0.041046143 0.1659288 0.16879799
#> [16,] 0.27265949 0.24691738 0.24163191 0.034857572 0.1540011 0.15531336
#> [17,] 0.47294022 0.33952493 0.15726696 0.024516755 0.2062132 0.22698193
#> [18,] 0.28614933 0.18189437 0.10744385 0.009640614 0.2522676 0.29974289
#> [19,] 0.62545203 0.37457313 0.36962881 0.032238939 0.5613243 0.68757940
#> [20,] 0.07144993 0.05825221 0.14829827 0.014518856 0.2144263 0.27809733
#> [21,] 0.65964784 0.57361635 0.59725734 0.012334206 0.6282340 0.49874924
#> [22,] 0.86501879 0.71800132 0.69691588 0.046998450 0.6991308 0.63073247
#> [23,] 0.47316123 0.27771473 0.33111362 0.075895733 0.4835435 0.40216936
#> [24,] 0.21134632 0.13953233 0.17291425 0.050140912 0.1852886 0.12177469
#> [25,] 0.40222326 0.30052370 0.21004987 0.020574148 0.1767231 0.17225305
#>              V42         V43        V44        V45        V46        V47
#>  [1,] 0.17435539  0.41024189 0.22341914 0.06743933 0.14176020 0.08620681
#>  [2,] 0.32912427 -0.01819204 0.11633414 0.19029421 0.19758068 0.10535533
#>  [3,] 0.11369120  0.18590782 0.19403748 0.17218684 0.16166485 0.14617700
#>  [4,] 0.46305643  0.39772574 0.41983800 0.51524204 0.44153609 0.26998531
#>  [5,] 0.10425099  0.02729446 0.06509431 0.04217272 0.02253118 0.02367357
#>  [6,] 0.22016506  0.16548880 0.06432182 0.06111254 0.04653377 0.03676585
#>  [7,] 0.32514684  0.21344713 0.06555162 0.11789430 0.15231895 0.17510844
#>  [8,] 0.08597935  0.11025562 0.06749325 0.06074221 0.05630883 0.09726177
#>  [9,] 0.21602231  0.09079444 0.20125464 0.20922648 0.13885958 0.07832930
#> [10,] 0.43000854  0.34689781 0.24656307 0.18635224 0.12971425 0.13621737
#> [11,] 0.29796676  0.23177076 0.17702710 0.15359603 0.14047207 0.10822209
#> [12,] 0.30368743  0.36282110 0.38584977 0.31162122 0.19889447 0.08088171
#> [13,] 0.28484359  0.37352377 0.43332662 0.41582425 0.20237557 0.12187504
#> [14,] 0.32965239  0.28099249 0.24184264 0.16946386 0.05843538 0.03913276
#> [15,] 0.06409175  0.05505090 0.03214402 0.04990401 0.02213542 0.01519786
#> [16,] 0.16678204  0.17106470 0.12395001 0.10390708 0.09167218 0.09141941
#> [17,] 0.18306576  0.12249864 0.08580812 0.08527504 0.06399478 0.03364696
#> [18,] 0.32936046  0.28988826 0.27887440 0.25962106 0.18278995 0.08395147
#> [19,] 0.66609647  0.66307443 0.47572311 0.30344815 0.19952357 0.12205993
#> [20,] 0.27354024  0.20920083 0.13038439 0.10258033 0.10431676 0.05248273
#> [21,] 0.44645061  0.30825903 0.14943456 0.09528196 0.02959264 0.07605586
#> [22,] 0.54921430  0.36360757 0.17564793 0.13159949 0.12794884 0.13174144
#> [23,] 0.25725306  0.21119654 0.20979123 0.15069264 0.10357656 0.09285741
#> [24,] 0.11539975  0.08781553 0.09074580 0.09142932 0.09260143 0.04837643
#> [25,] 0.13537620  0.17858638 0.14404459 0.09126402 0.06736903 0.04296951
#>              V48         V49          V5         V50         V51          V52
#>  [1,] 0.03093521 0.043243323 0.001286875 0.022258288 0.042469090 0.0144149977
#>  [2,] 0.06650391 0.014328331 0.132602878 0.019031920 0.016482117 0.0209463011
#>  [3,] 0.10065346 0.054829983 0.050894690 0.016772494 0.014932659 0.0125386969
#>  [4,] 0.18831139 0.098552165 0.069637278 0.036513764 0.032992960 0.0290609460
#>  [5,] 0.03403006 0.009193599 0.046348072 0.004028825 0.011613972 0.0066844619
#>  [6,] 0.02296958 0.020117358 0.064136504 0.010777550 0.011750391 0.0139841922
#>  [7,] 0.14396885 0.066774966 0.076996465 0.008166783 0.017903039 0.0148053482
#>  [8,] 0.10358607 0.063861182 0.151253117 0.028559967 0.026966195 0.0327371781
#>  [9,] 0.02753406 0.025994850 0.091207269 0.002712791 0.004526855 0.0071704404
#> [10,] 0.06727560 0.032884662 0.164763889 0.020078426 0.013175538 0.0139984243
#> [11,] 0.07267597 0.051145691 0.129792297 0.024521937 0.018613180 0.0195608041
#> [12,] 0.03287652 0.029011619 0.044809385 0.011342862 0.003281494 0.0090764199
#> [13,] 0.13358430 0.072466056 0.078941917 0.017719518 0.022700135 0.0150289991
#> [14,] 0.01082450 0.015955597 0.143517404 0.001520875 0.002910235 0.0025663451
#> [15,] 0.01566844 0.007041807 0.052069856 0.010755613 0.008402643 0.0072919347
#> [16,] 0.06918759 0.034725134 0.041266152 0.022229435 0.011516814 0.0152324828
#> [17,] 0.03323316 0.013619596 0.056690756 0.004816180 0.002227121 0.0037750257
#> [18,] 0.02903056 0.024190395 0.011461098 0.011226507 0.007154320 0.0090242153
#> [19,] 0.07574349 0.020944287 0.019025204 0.022329960 0.007699146 0.0009176766
#> [20,] 0.03739936 0.030033365 0.028184500 0.010301234 0.004206325 0.0090047227
#> [21,] 0.07418230 0.039468868 0.035794230 0.015612053 0.005975702 0.0053809958
#> [22,] 0.10825442 0.053096796 0.031647111 0.028677272 0.028577797 0.0058743723
#> [23,] 0.07892293 0.040236869 0.114746504 0.013208856 0.015299359 0.0126492278
#> [24,] 0.04918767 0.019480347 0.078785629 0.015970430 0.006116410 0.0211313543
#> [25,] 0.04271044 0.018855135 0.034994256 0.010735261 0.007643713 0.0075424752
#>                V53          V54           V55         V56          V57
#>  [1,]  0.007747753 0.0363517980  0.0002821613 0.008743586 0.0091505014
#>  [2,]  0.023151712 0.0323492370  0.0163576771 0.014073168 0.0133656521
#>  [3,]  0.006822705 0.0076646941  0.0040063000 0.005108660 0.0043335253
#>  [4,]  0.016487548 0.0168010753  0.0129838079 0.011714874 0.0114143948
#>  [5,] -0.003575445 0.0040659259  0.0021827684 0.004042967 0.0078766075
#>  [6,]  0.011269569 0.0210254333  0.0223581647 0.009497828 0.0039166643
#>  [7,]  0.008760684 0.0070814555  0.0038768742 0.006595050 0.0045660739
#>  [8,]  0.019029548 0.0111849575  0.0183210483 0.013196692 0.0061740029
#>  [9,]  0.002168356 0.0001704565 -0.0009478156 0.001347684 0.0008360507
#> [10,]  0.014246082 0.0078366636  0.0022120116 0.009567712 0.0092265238
#> [11,]  0.014669686 0.0176419905  0.0077191636 0.010694172 0.0090811024
#> [12,]  0.008386441 0.0037883247  0.0031037623 0.004153853 0.0018157706
#> [13,]  0.008751451 0.0117685889  0.0097491597 0.005009675 0.0042809359
#> [14,]  0.006520545 0.0209649315  0.0169931320 0.017050769 0.0164226811
#> [15,]  0.005303362 0.0018567165  0.0010129693 0.004735371 0.0036678876
#> [16,]  0.011253149 0.0099054635  0.0095406093 0.008208049 0.0059163112
#> [17,]  0.007927232 0.0035246357  0.0064338819 0.005620863 0.0027300130
#> [18,]  0.006591123 0.0071610060  0.0026864400 0.001125342 0.0012451001
#> [19,]  0.007649233 0.0069878762  0.0012996849 0.012690021 0.0131473205
#> [20,]  0.003244181 0.0033365195  0.0054873155 0.002921952 0.0029744144
#> [21,]  0.020122780 0.0155134074  0.0089173806 0.007659027 0.0080028948
#> [22,]  0.006363634 0.0071444695  0.0129927021 0.010578728 0.0025840945
#> [23,]  0.011262286 0.0126911630  0.0073432294 0.011985081 0.0117597551
#> [24,]  0.009834134 0.0136295979  0.0131823593 0.006762132 0.0090018083
#> [25,]  0.009350600 0.0081736560  0.0066518199 0.005713789 0.0067050300
#>               V58           V59         V6          V60         V7         V8
#>  [1,] 0.012656097  0.0137664124 0.12011873 0.0132126727 0.19733132 0.27055553
#>  [2,] 0.010771242 -0.0005565495 0.19197263 0.0029364080 0.16545494 0.24056199
#>  [3,] 0.004718063  0.0045705859 0.10366050 0.0065454683 0.13055198 0.14706238
#>  [4,] 0.011929271  0.0117832066 0.12435921 0.0075937584 0.13721195 0.21536692
#>  [5,] 0.004458344  0.0004343758 0.07313282 0.0007776183 0.14894429 0.14512289
#>  [6,] 0.010731598  0.0068691859 0.04894618 0.0027901169 0.09460722 0.11806292
#>  [7,] 0.006100265  0.0042542651 0.11741793 0.0033903390 0.09915890 0.09060586
#>  [8,] 0.015489426  0.0224444147 0.06728339 0.0147853001 0.09183255 0.04082019
#>  [9,] 0.005827745  0.0028579284 0.13495469 0.0039894808 0.17468860 0.20821924
#> [10,] 0.010458088  0.0154688128 0.04263813 0.0125114037 0.07802308 0.08342104
#> [11,] 0.009500610  0.0091953839 0.16893423 0.0063765096 0.17711594 0.20147723
#> [12,] 0.001246956  0.0031497718 0.05246192 0.0033284839 0.05933964 0.03867599
#> [13,] 0.005280759  0.0076023009 0.12302101 0.0049232915 0.13505176 0.19143805
#> [14,] 0.014070856  0.0171026902 0.13351274 0.0113828861 0.10844857 0.11123001
#> [15,] 0.001905966  0.0019253145 0.08164660 0.0016451024 0.09392618 0.09010716
#> [16,] 0.007745556  0.0049381533 0.05303011 0.0031786614 0.07545138 0.07295117
#> [17,] 0.004225887  0.0071884341 0.09501771 0.0072712231 0.10000691 0.07685796
#> [18,] 0.002528140  0.0063747345 0.01180872 0.0039128920 0.05109741 0.02327963
#> [19,] 0.006515130  0.0092940711 0.09385645 0.0070096061 0.08223321 0.09866222
#> [20,] 0.003477814  0.0008619030 0.04935055 0.0028975572 0.06882457 0.06972848
#> [21,] 0.008083597  0.0106011763 0.05314690 0.0090225096 0.02506902 0.04725752
#> [22,] 0.003297866  0.0073617951 0.07724152 0.0074539287 0.07630715 0.06120327
#> [23,] 0.007541810  0.0105699743 0.17590167 0.0079520161 0.17223594 0.17560492
#> [24,] 0.005840347  0.0067590931 0.06378916 0.0043195869 0.04879822 0.03730315
#> [25,] 0.004932572  0.0061499087 0.07039781 0.0043992908 0.06232162 0.05609798
#>                V9
#>  [1,]  0.25640350
#>  [2,]  0.37800691
#>  [3,]  0.20539939
#>  [4,]  0.34109990
#>  [5,]  0.23128165
#>  [6,]  0.08007975
#>  [7,]  0.14672670
#>  [8,]  0.14566456
#>  [9,]  0.23236136
#> [10,]  0.03750060
#> [11,]  0.23531547
#> [12,]  0.10817283
#> [13,]  0.31391805
#> [14,]  0.10804366
#> [15,]  0.16965393
#> [16,]  0.06011750
#> [17,]  0.10751953
#> [18,]  0.03362916
#> [19,]  0.05068584
#> [20,]  0.10954591
#> [21,]  0.04201883
#> [22,] -0.07771231
#> [23,]  0.19749580
#> [24,]  0.09023876
#> [25,]  0.10013128
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551