Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12         V13        V14
#>  [1,] 0.026295646 0.39567494 0.39424848 0.33180742 0.284609263 0.16059836
#>  [2,] 0.017249968 0.31624481 0.27884744 0.29360775 0.243693976 0.17692698
#>  [3,] 0.013443902 0.21068004 0.27429499 0.33409833 0.305702608 0.22792646
#>  [4,] 0.103735576 0.59051152 0.47787158 0.23995583 0.112955684 0.31492337
#>  [5,] 0.044600066 0.21408033 0.21360099 0.22985429 0.331994770 0.34410993
#>  [6,] 0.096800000 0.50250000 0.34880000 0.17000000 0.207600000 0.30870000
#>  [7,] 0.009400000 0.50790000 0.33500000 0.08340000 0.300400000 0.39570000
#>  [8,] 0.037466207 0.29364776 0.37137414 0.39536653 0.316048788 0.12725419
#>  [9,] 0.018735702 0.18864787 0.17561380 0.13986457 0.138026267 0.12977209
#> [10,] 0.047860848 0.24808421 0.33075728 0.36183897 0.453621397 0.58338283
#> [11,] 0.108300000 0.59660000 0.53040000 0.22510000 0.240200000 0.26890000
#> [12,] 0.022529731 0.31298136 0.37733374 0.38360476 0.366201830 0.29686242
#> [13,] 0.102100000 0.13690000 0.25090000 0.26310000 0.279600000 0.29770000
#> [14,] 0.010649689 0.14569310 0.24016614 0.26398174 0.226580848 0.16527818
#> [15,] 0.017755137 0.10242517 0.07312478 0.14734201 0.177679466 0.12972029
#> [16,] 0.004433460 0.09017170 0.09820535 0.02748329 0.001140945 0.05302893
#> [17,] 0.010000000 0.12640000 0.08810000 0.19920000 0.018400000 0.22610000
#> [18,] 0.021581197 0.12297340 0.11588813 0.10552634 0.192718105 0.34875312
#> [19,] 0.017306676 0.11684193 0.12018629 0.17809764 0.223538234 0.23582721
#> [20,] 0.035200000 0.29120000 0.23280000 0.22370000 0.247000000 0.15600000
#> [21,] 0.032852030 0.13207385 0.25657010 0.35334781 0.270018141 0.26058176
#> [22,] 0.013646099 0.09317510 0.05852860 0.09397892 0.240804968 0.27487661
#> [23,] 0.020859941 0.38904780 0.29461740 0.14387629 0.328822819 0.57253766
#> [24,] 0.009266751 0.09766237 0.09878090 0.11422933 0.137561511 0.14588571
#> [25,] 0.018011623 0.13128433 0.14945193 0.12325992 0.176876953 0.21731707
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.17264303 0.20159182 0.21213315 0.2671876 0.2337562 0.038069705
#>  [2,] 0.14932482 0.20087209 0.20485594 0.2308493 0.1846019 0.023957962
#>  [3,] 0.16492202 0.15917887 0.17468334 0.1905624 0.3041219 0.042756212
#>  [4,] 0.46793505 0.46454042 0.17465668 0.1384503 0.3608451 0.118096019
#>  [5,] 0.28544703 0.33846827 0.35447693 0.2370473 0.3445630 0.065252508
#>  [6,] 0.42240000 0.53120000 0.24360000 0.1884000 0.1908000 0.082100000
#>  [7,] 0.37690000 0.38280000 0.12470000 0.1363000 0.2678000 0.061100000
#>  [8,] 0.14328009 0.17635101 0.16781767 0.2311498 0.2278405 0.016131257
#>  [9,] 0.27979299 0.40019894 0.49633849 0.5148204 0.5884834 0.025604330
#> [10,] 0.69409357 0.72735824 0.77459427 0.8298590 0.9078787 0.051918577
#> [11,] 0.66460000 0.66320000 0.16740000 0.0837000 0.4331000 0.107000000
#> [12,] 0.24726099 0.30815657 0.37620889 0.4429265 0.5734987 0.029742517
#> [13,] 0.38230000 0.31290000 0.39560000 0.2093000 0.3218000 0.083000000
#> [14,] 0.07076026 0.13033246 0.29339269 0.4245629 0.5440827 0.008218980
#> [15,] 0.14827992 0.16156278 0.16541574 0.1854421 0.1707834 0.035201916
#> [16,] 0.13300059 0.09132067 0.09735997 0.1825063 0.2812827 0.002000377
#> [17,] 0.17290000 0.21310000 0.06930000 0.2281000 0.4060000 0.017100000
#> [18,] 0.58561491 0.78520603 0.88049484 0.8526099 0.8688079 0.015574401
#> [19,] 0.20144506 0.27901282 0.31905331 0.3398453 0.2872744 0.029535264
#> [20,] 0.34910000 0.33080000 0.22990000 0.2203000 0.2493000 0.011600000
#> [21,] 0.30047373 0.40456303 0.54670731 0.4760452 0.4689131 0.060671416
#> [22,] 0.17502452 0.17475747 0.12453730 0.3059783 0.6273766 0.030757663
#> [23,] 0.71017339 0.93699926 0.96595293 0.8972002 0.6184664 0.049010712
#> [24,] 0.18016738 0.17963526 0.15105268 0.1662073 0.1308904 0.012060442
#> [25,] 0.23597016 0.23777044 0.22918125 0.2699698 0.3047152 0.020932551
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.1715981 0.2881888 0.6233426 0.8003381 0.8275394 0.8469209 0.8671366
#>  [2,] 0.1301107 0.1043134 0.2272779 0.4188405 0.5724588 0.6962201 0.8197652
#>  [3,] 0.3122427 0.3977783 0.4762093 0.5293917 0.6250028 0.6575136 0.7326286
#>  [4,] 0.9007510 0.9418557 0.4657672 0.1289238 0.1373552 0.2665792 0.5332326
#>  [5,] 0.4494434 0.4911947 0.5493993 0.5816027 0.5775220 0.5624260 0.5998497
#>  [6,] 0.8321000 1.0000000 0.4076000 0.0960000 0.1928000 0.2419000 0.3790000
#>  [7,] 0.9188000 0.9779000 0.3236000 0.1944000 0.1874000 0.0885000 0.3443000
#>  [8,] 0.2610456 0.3316471 0.5876559 0.7883775 0.8670624 0.9860549 1.0668160
#>  [9,] 0.7816908 0.9379460 1.0182548 0.9142049 0.7355327 0.5381165 0.3021646
#> [10,] 0.9418303 0.9381550 0.8884897 0.7764506 0.6310052 0.4834396 0.4373489
#> [11,] 0.8718000 0.7992000 0.3712000 0.1703000 0.1611000 0.2086000 0.2847000
#> [12,] 0.6969579 0.8047422 0.8575200 0.8569364 0.8956007 0.9154056 0.9378446
#> [13,] 0.3345000 0.3184000 0.2887000 0.3610000 0.2566000 0.4106000 0.4591000
#> [14,] 0.6146365 0.6973118 0.8033988 0.8079083 0.7699415 0.8273677 0.9738343
#> [15,] 0.1679614 0.2075232 0.1618556 0.1683562 0.1998973 0.1968665 0.2856100
#> [16,] 0.3351654 0.3547538 0.4928554 0.6796489 0.8648877 0.9498711 0.8077219
#> [17,] 0.3973000 0.2741000 0.3690000 0.5556000 0.4846000 0.3140000 0.5334000
#> [18,] 0.8528978 0.9063639 0.9122073 0.8604721 0.8616001 0.8787303 0.9580187
#> [19,] 0.4153260 0.5399676 0.5706487 0.6201854 0.5803078 0.7035261 0.8286025
#> [20,] 0.4128000 0.3158000 0.6191000 0.5854000 0.3395000 0.2561000 0.5599000
#> [21,] 0.6667904 0.4592855 0.2849705 0.4287875 0.8860413 0.8490196 0.4223680
#> [22,] 0.8278617 0.8194927 0.8494771 0.5366713 0.3574778 0.5471510 0.8259816
#> [23,] 0.4475802 0.4769673 0.5276140 0.7144309 0.7991563 0.7615422 0.7320405
#> [24,] 0.1465100 0.2245966 0.2266141 0.3631892 0.5967044 0.8546565 0.9931514
#> [25,] 0.3547960 0.4012093 0.4731754 0.5733027 0.6855915 0.7166416 0.7708168
#>             V27          V28       V29          V3       V30       V31
#>  [1,] 0.9390889  1.009780342 0.7938036  0.04426509 0.5438727 0.6448976
#>  [2,] 0.9076949  0.927120227 0.8982477  0.02198085 0.9336554 0.9140058
#>  [3,] 0.9244623  1.025413561 0.9095315  0.04273131 0.6762712 0.5830375
#>  [4,] 0.3416924  0.598740717 0.5712684  0.08284504 0.8447148 0.5415029
#>  [5,] 0.6544241  0.768046627 0.7717967  0.07894003 0.8125636 0.8252307
#>  [6,] 0.2893000  0.345100000 0.3777000  0.06290000 0.5213000 0.2316000
#>  [7,] 0.2953000  0.590800000 0.4564000  0.11360000 0.7334000 0.1969000
#>  [8,] 0.9802486  0.956516073 0.9136096  0.02427009 0.8043866 0.8106732
#>  [9,] 0.1103452 -0.001120043 0.2414165  0.01194302 0.3464535 0.2803635
#> [10,] 0.4285691  0.366179909 0.3159661  0.04809495 0.3181557 0.2715379
#> [11,] 0.2211000  0.613400000 0.5807000  0.02570000 0.6925000 0.3825000
#> [12,] 0.9300328  0.858158979 0.7134070  0.03475223 0.5567957 0.4396443
#> [13,] 0.4722000  0.727800000 0.7591000  0.05770000 0.6579000 0.7514000
#> [14,] 1.0401567  1.005416715 0.8849726  0.03658310 0.6617834 0.2648002
#> [15,] 0.3666503  0.567958725 0.6438547  0.02639178 0.6513950 0.6493800
#> [16,] 0.7759288  0.784827697 0.6050649 -0.02059643 0.4083080 0.2275250
#> [17,] 0.5256000  0.252000000 0.2090000  0.06230000 0.3559000 0.6260000
#> [18,] 0.8692128  0.692118018 0.5150535  0.02584002 0.4662938 0.4675494
#> [19,] 0.8054585  0.628249656 0.4287532  0.05296644 0.3879065 0.4737495
#> [20,] 0.8145000  0.694100000 0.6985000  0.01910000 0.8660000 0.5930000
#> [21,] 0.2837053  0.618692498 0.7508577  0.06865117 0.4523784 0.5532452
#> [22,] 1.0187333  0.859129629 0.6154002  0.02174213 0.4188512 0.3944722
#> [23,] 0.7544140  0.590071762 0.2740118  0.05709889 0.3190270 0.5877316
#> [24,] 0.8374526  0.639001121 0.5339144  0.01130477 0.3876243 0.3082546
#> [25,] 0.7564619  0.829573000 0.8779300  0.02498281 0.8282602 0.7424595
#>             V32       V33       V34         V35        V36       V37        V38
#>  [1,] 0.6421990 0.4913661 0.2105178  0.09607267 0.04118028 0.1603586 0.19267748
#>  [2,] 0.8265015 0.7137710 0.6496361  0.65907073 0.70907031 0.6910770 0.57261199
#>  [3,] 0.5344407 0.4123318 0.3861185  0.20813085 0.04577794 0.1989499 0.14867465
#>  [4,] 0.5153750 0.5199865 0.6557205  0.84062482 0.81966251 0.8188192 0.89680922
#>  [5,] 0.7702939 0.5971529 0.4727649  0.40876434 0.25580338 0.2015509 0.32003567
#>  [6,] 0.3335000 0.4781000 0.6116000  0.67050000 0.73750000 0.7356000 0.77920000
#>  [7,] 0.2790000 0.6212000 0.8681000  0.86210000 0.93800000 0.8327000 0.94800000
#>  [8,] 0.6516400 0.4729255 0.2684682  0.13859728 0.12574987 0.1376489 0.13059314
#>  [9,] 0.4658938 0.3863325 0.1535015 -0.03771102 0.20476459 0.4049005 0.34850825
#> [10,] 0.3165614 0.3538909 0.3029390  0.28414556 0.18401286 0.2600123 0.32908143
#> [11,] 0.4303000 0.7791000 0.8703000  1.00000000 0.92120000 0.9386000 0.93030000
#> [12,] 0.4085628 0.4029557 0.3077607  0.19647248 0.23181114 0.2570669 0.26888585
#> [13,] 0.6666000 0.4903000 0.5962000  0.65520000 0.40140000 0.1188000 0.32450000
#> [14,] 0.2042342 0.2873469 0.4667316  0.71156101 0.84911075 0.7939726 0.64522812
#> [15,] 0.6050963 0.6825802 0.7658106  0.81368757 0.85618158 0.7747311 0.68554329
#> [16,] 0.2581584 0.2398360 0.2718370  0.19952866 0.20118803 0.2972584 0.31061696
#> [17,] 0.7340000 0.6120000 0.3497000  0.39530000 0.30120000 0.5408000 0.88140000
#> [18,] 0.3703308 0.3334120 0.3279650  0.21755810 0.11064689 0.2356564 0.17133131
#> [19,] 0.5015522 0.4213747 0.4124700  0.52618241 0.58403354 0.5307444 0.38879646
#> [20,] 0.3664000 0.6750000 0.8697000  0.78370000 0.75520000 0.5789000 0.47130000
#> [21,] 0.6995945 0.6048959 0.8721737  0.75804599 0.69622625 0.7513500 0.49266362
#> [22,] 0.2877652 0.2913198 0.2683349  0.28438357 0.26618302 0.3865205 0.43785824
#> [23,] 0.5929514 0.2866819 0.3771173  0.73245560 0.67261276 0.4953368 0.53638961
#> [24,] 0.1734386 0.1766190 0.2077880  0.27651605 0.20826204 0.0711585 0.07575517
#> [25,] 0.6441851 0.6308673 0.6385796  0.59857439 0.51397015 0.3696220 0.28232153
#>              V39          V4        V40        V41        V42        V43
#>  [1,] 0.18985251  0.07693758 0.11530371 0.16571294 0.06260217 0.08128750
#>  [2,] 0.48707568  0.04851324 0.50765535 0.48156360 0.42644153 0.44735341
#>  [3,] 0.18059319  0.05130201 0.36555614 0.39535010 0.26959721 0.16847149
#>  [4,] 0.76939659  0.13063975 0.50777791 0.38650561 0.45131863 0.45676457
#>  [5,] 0.35747339  0.07352632 0.31508868 0.55748353 0.60299501 0.45448357
#>  [6,] 0.67880000  0.06080000 0.52590000 0.27620000 0.15450000 0.20190000
#>  [7,] 0.67210000  0.12030000 0.44360000 0.51630000 0.38090000 0.15570000
#>  [8,] 0.16711501  0.05970511 0.04067805 0.15714418 0.10962026 0.15664478
#>  [9,] 0.24342781  0.04327671 0.14549837 0.14757045 0.07562895 0.02590127
#> [10,] 0.28945473  0.06101893 0.26445371 0.18156112 0.24399885 0.24423738
#> [11,] 0.73140000  0.08370000 0.47910000 0.20870000 0.20160000 0.16690000
#> [12,] 0.32265283  0.04599910 0.32071234 0.24661882 0.26969304 0.29417532
#> [13,] 0.31070000  0.06270000 0.13540000 0.51090000 0.79880000 0.75170000
#> [14,] 0.45763525  0.06056638 0.34421156 0.37831954 0.30601795 0.36779892
#> [15,] 0.65364658  0.02910657 0.61409368 0.52925253 0.47026287 0.33722950
#> [16,] 0.32107432 -0.02406479 0.27836747 0.23850268 0.17444498 0.13566563
#> [17,] 0.98570000  0.02050000 0.91670000 0.61210000 0.50060000 0.32100000
#> [18,] 0.08144868  0.03582940 0.12297441 0.07237518 0.05435226 0.10888423
#> [19,] 0.39840848  0.05255529 0.50202224 0.44814923 0.37270581 0.31821514
#> [20,] 0.12520000  0.04690000 0.60870000 0.73220000 0.59770000 0.34310000
#> [21,] 0.35456387  0.09836571 0.34873273 0.29787819 0.26995189 0.06800958
#> [22,] 0.36402901  0.03125115 0.25951369 0.18115094 0.13851242 0.07976922
#> [23,] 0.43648769  0.04645588 0.29803945 0.13191769 0.16604642 0.25149636
#> [24,] 0.13438870  0.02775881 0.20905995 0.27523512 0.26268285 0.19811914
#> [25,] 0.23614713  0.02651629 0.23935323 0.22148545 0.21517420 0.24384829
#>              V44        V45        V46        V47        V48         V49
#>  [1,] 0.13743542 0.06831188 0.06397494 0.07665925 0.11250914 0.078013204
#>  [2,] 0.51745804 0.47915164 0.33801098 0.23782490 0.19303431 0.105293086
#>  [3,] 0.09713780 0.14660714 0.16096196 0.15986355 0.15485527 0.076650326
#>  [4,] 0.43603107 0.54591388 0.42258046 0.09149194 0.09083744 0.035948340
#>  [5,] 0.44372497 0.57986423 0.53622511 0.33561079 0.19610285 0.103322169
#>  [6,] 0.22310000 0.42210000 0.30670000 0.13290000 0.13490000 0.105700000
#>  [7,] 0.14490000 0.26620000 0.18060000 0.16990000 0.25590000 0.112900000
#>  [8,] 0.18435808 0.12039370 0.03577592 0.08117393 0.10671253 0.064817621
#>  [9,] 0.06045879 0.04076455 0.02522976 0.03732401 0.03432217 0.002974037
#> [10,] 0.14407874 0.10097073 0.08021399 0.07023673 0.06151167 0.045991046
#> [11,] 0.28720000 0.43740000 0.30970000 0.15780000 0.05530000 0.033400000
#> [12,] 0.27914099 0.25175162 0.14837718 0.12115834 0.11411400 0.063393111
#> [13,] 0.55080000 0.58580000 0.72920000 0.55220000 0.33390000 0.160800000
#> [14,] 0.36683615 0.30742982 0.19760572 0.07255515 0.01403425 0.027570981
#> [15,] 0.21735185 0.21163175 0.19247897 0.17188057 0.13658384 0.077492206
#> [16,] 0.20686540 0.18528646 0.12959379 0.02668799 0.02123628 0.019337688
#> [17,] 0.32020000 0.42950000 0.36540000 0.26550000 0.15760000 0.068100000
#> [18,] 0.10871210 0.12014435 0.07730218 0.08605603 0.06839971 0.038995499
#> [19,] 0.22281154 0.13010965 0.10518854 0.05844781 0.04084302 0.019221464
#> [20,] 0.18030000 0.23780000 0.34240000 0.23030000 0.06890000 0.021600000
#> [21,] 0.19951091 0.16312060 0.16491670 0.11472818 0.06531625 0.044399568
#> [22,] 0.04427201 0.08205807 0.07835616 0.08767474 0.05385791 0.020623853
#> [23,] 0.25574151 0.20163984 0.08750121 0.07007469 0.05543586 0.031011001
#> [24,] 0.09943180 0.09190941 0.09638815 0.04300871 0.03823478 0.031324598
#> [25,] 0.20703647 0.13806362 0.10116927 0.06287315 0.04402355 0.024837268
#>                 V5         V50           V51           V52          V53
#>  [1,]  0.104393371 0.028011526  0.0156049482  0.0095013813  0.004670256
#>  [2,]  0.077387185 0.029951872  0.0175446283  0.0143734333  0.010242080
#>  [3,]  0.053408608 0.004462564  0.0212186621  0.0194739474  0.011246015
#>  [4,]  0.128517714 0.020753334  0.0149007197  0.0428318302  0.024308064
#>  [5,]  0.099591452 0.045418539  0.0336812826  0.0290084982  0.020345312
#>  [6,]  0.061700000 0.049900000  0.0206000000  0.0073000000  0.008100000
#>  [7,]  0.040300000 0.020100000  0.0480000000  0.0234000000  0.017500000
#>  [8,]  0.072004806 0.013563650  0.0079920075  0.0081408666  0.006695014
#>  [9,]  0.044395040 0.005546332  0.0125037505  0.0084329350 -0.001187334
#> [10,]  0.087038684 0.024038942  0.0204032381  0.0201060460  0.015844940
#> [11,]  0.074800000 0.020900000  0.0172000000  0.0180000000  0.011000000
#> [12,]  0.066103529 0.016259825  0.0204672642  0.0125961646  0.008640771
#> [13,]  0.063500000 0.047500000  0.1004000000  0.0709000000  0.031700000
#> [14,]  0.033142200 0.007867613 -0.0007440928  0.0102735045  0.009173790
#> [15,]  0.039002566 0.028994308  0.0197053056  0.0099067976  0.014594273
#> [16,] -0.009146163 0.004207379 -0.0051532093 -0.0002705236  0.004662220
#> [17,]  0.020500000 0.029400000  0.0241000000  0.0121000000  0.003600000
#> [18,]  0.052669060 0.017425258  0.0147956948  0.0083481524  0.008450377
#> [19,]  0.068786737 0.010755992  0.0094299406  0.0059603742  0.008272201
#> [20,]  0.073700000 0.046900000  0.0426000000  0.0346000000  0.015800000
#> [21,]  0.166177397 0.007992483  0.0145934424  0.0057042781  0.014890313
#> [22,]  0.038848742 0.024334679  0.0165484826  0.0141267894  0.025259702
#> [23,]  0.095879508 0.026019879  0.0056635625  0.0184436296  0.005605607
#> [24,]  0.047204506 0.007894537  0.0081125630  0.0103163020  0.005005744
#> [25,]  0.046261117 0.013963559  0.0107126507  0.0097998205  0.007737960
#>               V54          V55          V56           V57          V58
#>  [1,] 0.006444598  0.003946819  0.009906902  0.0055829306 0.0062076130
#>  [2,] 0.007335984  0.005639415  0.007202900  0.0062204597 0.0046532927
#>  [3,] 0.008822448  0.008226652  0.008090154  0.0030648405 0.0043657005
#>  [4,] 0.011802879  0.013266730  0.018093566  0.0109457991 0.0182566980
#>  [5,] 0.017181050  0.016855186  0.014060494  0.0135052536 0.0148738483
#>  [6,] 0.030300000  0.019000000  0.021200000  0.0126000000 0.0201000000
#>  [7,] 0.035200000  0.015800000  0.032600000  0.0201000000 0.0168000000
#>  [8,] 0.008200807  0.005082074  0.005952027  0.0042004699 0.0034731065
#>  [9,] 0.003996219  0.003533705  0.002342603  0.0076275245 0.0041607349
#> [10,] 0.018450612  0.017519391  0.012336955  0.0044146679 0.0168002015
#> [11,] 0.023400000  0.027600000  0.003200000  0.0084000000 0.0122000000
#> [12,] 0.009468970  0.007923000  0.006603495  0.0053485781 0.0061954777
#> [13,] 0.030900000  0.025200000  0.008700000  0.0177000000 0.0214000000
#> [14,] 0.001193330  0.002658239  0.002688248 -0.0013199808 0.0003703622
#> [15,] 0.011017659  0.014228187  0.009399014  0.0059435902 0.0107368320
#> [16,] 0.007113338 -0.001911371 -0.000745409 -0.0002895302 0.0030724212
#> [17,] 0.015000000  0.008500000  0.007300000  0.0050000000 0.0044000000
#> [18,] 0.006838720  0.004600696  0.007390021  0.0054702389 0.0028341309
#> [19,] 0.009494633  0.009297842  0.008835090  0.0094614402 0.0066286026
#> [20,] 0.015400000  0.010900000  0.004800000  0.0095000000 0.0015000000
#> [21,] 0.007078928  0.018597285  0.007263641  0.0137136647 0.0054147972
#> [22,] 0.021141018  0.006928142  0.010502786  0.0032262658 0.0092306242
#> [23,] 0.016763043  0.010366297  0.003365995  0.0060886026 0.0046992584
#> [24,] 0.004424893  0.005253904  0.001931799  0.0024255015 0.0038303976
#> [25,] 0.007147593  0.005752535  0.006450207  0.0067077987 0.0054789011
#>                V59          V6          V60         V7         V8         V9
#>  [1,] 0.0087119616  0.12142132  0.011541487 0.14822661 0.23116289 0.30935275
#>  [2,] 0.0027932270  0.10630896  0.003049523 0.13730422 0.14651395 0.27790279
#>  [3,] 0.0047421126  0.08530376  0.003995999 0.08534855 0.06946975 0.15369658
#>  [4,] 0.0151998091  0.14752296  0.017238455 0.17479722 0.39928038 0.66301816
#>  [5,] 0.0139571455  0.11847319  0.008204921 0.10749729 0.12572675 0.22557788
#>  [6,] 0.0210000000  0.12070000  0.004100000 0.09440000 0.42230000 0.57440000
#>  [7,] 0.0245000000  0.12270000  0.015400000 0.24950000 0.45660000 0.65870000
#>  [8,] 0.0061368796  0.08362452  0.008538326 0.08848603 0.13701329 0.14674647
#>  [9,] 0.0014087634  0.07975337 -0.000123733 0.14792370 0.14533649 0.19959658
#> [10,] 0.0104666990  0.09594099  0.009123525 0.12842885 0.11999171 0.18136175
#> [11,] 0.0082000000  0.11250000  0.014300000 0.33220000 0.45900000 0.55260000
#> [12,] 0.0063467992  0.09229981  0.004367546 0.10690672 0.16088485 0.24567924
#> [13,] 0.0227000000  0.13280000  0.010600000 0.09880000 0.17870000 0.11990000
#> [14,] 0.0027013629  0.03541626  0.003132723 0.03804231 0.03207121 0.09779100
#> [15,] 0.0122657754  0.07190164  0.007668700 0.09880225 0.09985173 0.03988687
#> [16,] 0.0006751207 -0.03221422  0.004463272 0.02995771 0.08200958 0.07328687
#> [17,] 0.0040000000  0.03680000  0.011700000 0.10980000 0.12760000 0.05980000
#> [18,] 0.0048572099  0.08173597  0.003409871 0.12776424 0.09518078 0.12759514
#> [19,] 0.0085792473  0.10647207  0.008197031 0.08782387 0.09024177 0.07709565
#> [20,] 0.0073000000  0.11850000  0.006700000 0.16830000 0.15410000 0.14660000
#> [21,] 0.0173451023  0.24482033  0.006592196 0.23381637 0.20913226 0.18415227
#> [22,] 0.0014582247  0.01792317  0.003659789 0.02433694 0.04601618 0.10147589
#> [23,] 0.0108958612  0.21753685  0.007859004 0.21613350 0.21094587 0.31254779
#> [24,] 0.0024177645  0.08765349  0.002091491 0.09563486 0.08306189 0.11477688
#> [25,] 0.0078072225  0.09353573  0.006278313 0.10968498 0.10911071 0.11944929
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406