Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1          V10        V11        V12        V13        V14
#>  [1,] 0.020151421  0.173265398 0.10720090 0.03583164 0.09084100 0.23061854
#>  [2,] 0.040280826  0.387821039 0.41094323 0.37918269 0.30052652 0.12948663
#>  [3,] 0.032953222  0.356311569 0.44619597 0.49749401 0.50341747 0.41942862
#>  [4,] 0.023826212  0.158667973 0.29544596 0.31890130 0.28149266 0.29765975
#>  [5,] 0.027272246  0.245165228 0.30518800 0.36643407 0.25864504 0.16193413
#>  [6,] 0.054569275 -0.005677047 0.11557060 0.21433738 0.28740767 0.36115831
#>  [7,] 0.020888051  0.382777061 0.46439535 0.47818820 0.46956597 0.36786456
#>  [8,] 0.055196440  0.644829247 0.61069987 0.37367097 0.27984157 0.24290531
#>  [9,] 0.021692532  0.233876274 0.26075096 0.26308085 0.33933244 0.47857273
#> [10,] 0.033141802  0.269813124 0.23280295 0.26855899 0.26181341 0.20529567
#> [11,] 0.082123285  0.247064419 0.42859924 0.52255747 0.60625441 0.62406484
#> [12,] 0.028686596  0.201524287 0.27899920 0.33553559 0.33153455 0.34309982
#> [13,] 0.005994710  0.262559346 0.28640475 0.36054760 0.32858330 0.26300123
#> [14,] 0.024702687  0.126776737 0.14492239 0.19460315 0.26063935 0.37840656
#> [15,] 0.026138245  0.097088708 0.14634754 0.05377482 0.09293284 0.10577311
#> [16,] 0.013694730  0.129752957 0.07904832 0.07944451 0.07280551 0.09836018
#> [17,] 0.014649557  0.104701842 0.08350290 0.16334504 0.18241236 0.14298406
#> [18,] 0.023171512  0.137481867 0.09192956 0.02609689 0.14507623 0.28893933
#> [19,] 0.018941620  0.175689017 0.13855611 0.13128298 0.14092819 0.14223210
#> [20,] 0.018575287  0.114053997 0.13879838 0.12385064 0.13016380 0.09414200
#> [21,] 0.015258837  0.107158468 0.12083372 0.11282043 0.12083488 0.14776858
#> [22,] 0.003359666  0.096951986 0.10180804 0.24078950 0.28930635 0.28889025
#> [23,] 0.066807304  0.307264506 0.28042364 0.08141967 0.14201439 0.47690563
#> [24,] 0.004509667  0.212057504 0.13391789 0.08136971 0.22238211 0.28937398
#> [25,] 0.018171141  0.168322658 0.21570153 0.09046007 0.19043825 0.41359479
#>             V15        V16        V17       V18       V19           V2
#>  [1,] 0.3968781 0.59090975 0.72560580 0.7754973 0.8556858  0.017587004
#>  [2,] 0.1494032 0.19604422 0.17434274 0.2583903 0.2640153  0.059209212
#>  [3,] 0.3158375 0.28556528 0.28748097 0.2858599 0.4378983  0.047776456
#>  [4,] 0.2624119 0.15777305 0.22243553 0.2738154 0.2721252  0.040449211
#>  [5,] 0.1444056 0.16019550 0.09019324 0.1741775 0.1783173  0.006012075
#>  [6,] 0.4631969 0.49949160 0.66240854 0.7948019 0.9046570  0.050153507
#>  [7,] 0.2455737 0.22427723 0.29614176 0.3636684 0.4593037  0.009939480
#>  [8,] 0.3611660 0.43547091 0.22371478 0.2062457 0.3190395  0.088192768
#>  [9,] 0.5223379 0.54819420 0.63164998 0.7214031 0.7473425  0.016416622
#> [10,] 0.1640787 0.20628376 0.22290430 0.2148293 0.2180683  0.031785797
#> [11,] 0.6802845 0.80682488 0.93859065 0.9318169 0.9099504  0.098757471
#> [12,] 0.4079255 0.49904548 0.61162526 0.6947745 0.7447032  0.031524172
#> [13,] 0.1754659 0.18315120 0.21384304 0.3525331 0.6239375  0.016825525
#> [14,] 0.5012834 0.60752659 0.71859372 0.7724167 0.7971718  0.028814098
#> [15,] 0.1436849 0.15809432 0.15232534 0.1445900 0.1153110  0.021137973
#> [16,] 0.0722763 0.12195730 0.24778672 0.2536628 0.3246307  0.015209324
#> [17,] 0.1679483 0.21219421 0.18059947 0.2168986 0.2616413  0.034460690
#> [18,] 0.5711695 0.82179490 0.86044205 0.8116430 0.8960535  0.017226447
#> [19,] 0.1827908 0.22432084 0.23567495 0.3457111 0.3797860  0.014234630
#> [20,] 0.0366881 0.05647537 0.09640437 0.1930386 0.3880444  0.008477299
#> [21,] 0.1864847 0.18504327 0.16510498 0.1671477 0.1728546  0.014835768
#> [22,] 0.1365125 0.20093179 0.30511166 0.3664405 0.3859995 -0.003409555
#> [23,] 0.6624714 0.68488039 0.73935314 0.6013173 0.3993947  0.060982507
#> [24,] 0.3798050 0.24616952 0.07110549 0.1397296 0.2116987  0.020127545
#> [25,] 0.5094711 0.73410834 0.90321606 0.9248119 0.7561035  0.013432616
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.9069661 0.9009873 0.9083286 0.8116361 0.5302826 0.3251240 0.3625051
#>  [2,] 0.1732682 0.2768154 0.6416491 0.8336691 0.8643204 0.8763961 0.8870287
#>  [3,] 0.6995612 0.8183686 0.7917602 0.6596517 0.6518711 0.6490566 0.8239092
#>  [4,] 0.2715514 0.3592744 0.3292749 0.4690792 0.6264747 0.6700732 0.6567263
#>  [5,] 0.2498471 0.3407378 0.2958279 0.4592208 0.6449896 0.7591108 0.8431342
#>  [6,] 0.9215889 0.8806834 0.9174376 0.9803326 0.9319973 0.7532072 0.5905216
#>  [7,] 0.6026047 0.7359670 0.8882511 0.8987154 0.8995583 0.9426959 0.9619467
#>  [8,] 0.6345030 0.7548363 0.6632811 0.6598488 0.6862632 0.4819317 0.4564934
#>  [9,] 0.7769129 0.8699698 0.9445045 0.9446792 0.9555641 0.9475039 0.9152837
#> [10,] 0.1896196 0.1479962 0.2585293 0.4513591 0.5390227 0.6181102 0.6855775
#> [11,] 0.8359237 0.7964717 0.6272819 0.4458796 0.3985533 0.3474704 0.2637818
#> [12,] 0.8206256 0.8773455 0.8971923 0.9039157 0.9773738 0.9858964 0.9460381
#> [13,] 0.7371950 0.8640454 0.8569311 0.8351456 0.9471351 0.9630009 0.9949765
#> [14,] 0.8153807 0.8796657 0.8809367 0.8336681 0.8184154 0.8065250 0.8036626
#> [15,] 0.1988261 0.3537708 0.6311060 0.8855458 0.8925270 0.8258257 0.8408005
#> [16,] 0.4759800 0.6070614 0.8031755 0.9367186 0.8913579 0.8208643 0.8856098
#> [17,] 0.1629441 0.2534820 0.2377850 0.2548837 0.2989807 0.2636176 0.3598778
#> [18,] 0.8991462 0.9884426 0.9863183 0.9780841 0.9808056 0.9149972 0.9642291
#> [19,] 0.5001566 0.4901387 0.5637786 0.4988901 0.4227502 0.3441396 0.4919983
#> [20,] 0.3744888 0.2650513 0.3070136 0.4714763 0.6424669 0.6558528 0.6419078
#> [21,] 0.2268932 0.2869191 0.3273037 0.4954681 0.6698266 0.8365583 0.9240784
#> [22,] 0.5496881 0.5686143 0.4463601 0.4532102 0.5825577 0.8169266 0.8265236
#> [23,] 0.2954964 0.5193126 0.7605891 0.8588024 0.9569782 0.9146727 0.6574366
#> [24,] 0.2242448 0.2173102 0.3680579 0.5914406 0.7621882 0.6903527 0.6998327
#> [25,] 0.4667742 0.5367450 0.7248637 0.8032565 0.8415234 0.7231032 0.6449209
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.3464281 0.4330232 0.5672942 0.010221629 0.6076265 0.4354355 0.4180763
#>  [2,] 0.8525097 0.8654944 0.7926860 0.072786993 0.6037317 0.6914231 0.6171160
#>  [3,] 0.9513701 0.9174829 0.8222725 0.058349071 0.6102670 0.5059220 0.5387104
#>  [4,] 0.8103062 0.9168309 0.8668694 0.052758756 0.6350349 0.5638727 0.4810981
#>  [5,] 0.9172789 1.0227791 0.9858797 0.045947757 0.7661544 0.6790004 0.4743746
#>  [6,] 0.5343847 0.5348143 0.4701463 0.080459993 0.3641344 0.2600299 0.2296438
#>  [7,] 0.9477535 0.8005856 0.5687360 0.017814673 0.4529916 0.4053526 0.4825423
#>  [8,] 0.3502761 0.4103569 0.4436826 0.076078986 0.5058641 0.3770610 0.2066363
#>  [9,] 0.8645167 0.7592012 0.6619187 0.027406436 0.5362574 0.3820410 0.2684418
#> [10,] 0.7728938 0.8812872 0.8890124 0.041402448 0.9012186 0.8957358 0.7948261
#> [11,] 0.2546087 0.3062915 0.2387120 0.111383625 0.2010497 0.2048988 0.2425490
#> [12,] 0.9039590 0.9257624 0.7901768 0.022096971 0.5817462 0.4428808 0.2852304
#> [13,] 1.0108723 1.0098596 0.9183000 0.030427786 0.6465848 0.3919272 0.3019157
#> [14,] 0.6692629 0.5142645 0.4910180 0.034905177 0.4974970 0.4121120 0.2719209
#> [15,] 0.7893300 0.8296769 0.9296990 0.028050352 0.7910345 0.6230724 0.5312488
#> [16,] 0.9896041 0.9603853 0.9181253 0.016810910 0.8111646 0.7602381 0.7086183
#> [17,] 0.3562911 0.4938947 0.6059280 0.034634345 0.6351296 0.6475293 0.5923127
#> [18,] 0.9611208 0.7570783 0.4431440 0.043823871 0.3708198 0.4615561 0.4125921
#> [19,] 0.7077983 0.8064807 0.8419545 0.007722611 0.7814147 0.6182623 0.4450281
#> [20,] 0.5631042 0.7287067 0.9859123 0.012339736 1.0022051 0.9584204 0.8987107
#> [21,] 0.8158739 0.6731831 0.5526981 0.016494922 0.4100805 0.3020150 0.1861561
#> [22,] 0.6906618 0.5727874 0.5066038 0.042996224 0.2856069 0.3303901 0.4731239
#> [23,] 0.6484031 0.6401299 0.4834179 0.062273725 0.5545862 0.4672032 0.2026262
#> [24,] 0.7081356 0.8595042 0.9942835 0.045587139 0.9504829 0.8779808 0.7599280
#> [25,] 0.5648330 0.3756636 0.2948554 0.039695164 0.3890470 0.2458313 0.2362697
#>             V33        V34        V35         V36        V37       V38
#>  [1,] 0.4804581 0.38129137 0.27853819  0.21456613 0.16722349 0.2149606
#>  [2,] 0.4569082 0.21340249 0.08479933  0.10685899 0.25220050 0.1828446
#>  [3,] 0.4582185 0.27156760 0.06595109  0.21583903 0.23214081 0.3611195
#>  [4,] 0.4679866 0.43458824 0.20710385  0.07200958 0.14292892 0.1162392
#>  [5,] 0.4284922 0.36838478 0.16343619 -0.02015073 0.15988601 0.1562438
#>  [6,] 0.2757875 0.32054124 0.37302387  0.38213961 0.38512237 0.4697067
#>  [7,] 0.4732839 0.38422510 0.22248872  0.37365001 0.45339569 0.3559117
#>  [8,] 0.5091638 0.70416686 0.70566690  0.60076230 0.57517488 0.5633079
#>  [9,] 0.1695994 0.05008388 0.08888377  0.16118160 0.21810028 0.2592295
#> [10,] 0.6718641 0.60284056 0.57677467  0.54136802 0.49840868 0.4791788
#> [11,] 0.2754486 0.30292262 0.36489153  0.15472297 0.15608087 0.2475582
#> [12,] 0.2916599 0.19191334 0.12094581  0.01136071 0.09689592 0.2430396
#> [13,] 0.2236686 0.20053527 0.18194601  0.06405580 0.03289688 0.0710084
#> [14,] 0.2731047 0.31456902 0.21316142  0.12293693 0.14880079 0.1450755
#> [15,] 0.6032552 0.74654410 0.73570357  0.56078687 0.30467707 0.1771635
#> [16,] 0.6616895 0.68311753 0.67221883  0.57040897 0.51293343 0.4277588
#> [17,] 0.6081030 0.63911777 0.70325066  0.75419498 0.74411654 0.7787806
#> [18,] 0.3508978 0.31381352 0.22686626  0.08540865 0.26622698 0.1960258
#> [19,] 0.5308826 0.55314271 0.47428401  0.50543695 0.48909618 0.4306346
#> [20,] 0.8264357 0.74967169 0.62307052  0.49238045 0.43402455 0.1875476
#> [21,] 0.1721302 0.21470195 0.24979472  0.22153456 0.13693226 0.1211597
#> [22,] 0.5616224 0.56048224 0.59423494  0.57458171 0.50854343 0.2578209
#> [23,] 0.2417406 0.15877232 0.49295565  0.72305326 0.74968087 0.3943795
#> [24,] 0.7046824 0.57421817 0.45986327  0.46499251 0.31450620 0.2098791
#> [25,] 0.2622844 0.11605657 0.36067241  0.59299833 0.55082848 0.3165996
#>              V39         V4        V40        V41        V42        V43
#>  [1,] 0.19495301 0.04025978 0.13483338 0.15530141 0.22398280 0.19859525
#>  [2,] 0.16309230 0.10784430 0.12419688 0.18260457 0.14320141 0.13573448
#>  [3,] 0.55779530 0.04139851 0.46302506 0.26445607 0.43939464 0.42301427
#>  [4,] 0.23882896 0.10620483 0.33588117 0.29893107 0.32402739 0.14288794
#>  [5,] 0.21094656 0.01880278 0.18103795 0.26473879 0.32755851 0.22402095
#>  [6,] 0.45442231 0.10246489 0.35874170 0.36160371 0.40528216 0.34509051
#>  [7,] 0.35705299 0.02754582 0.34447969 0.16512217 0.14198016 0.29549404
#>  [8,] 0.54854800 0.08347089 0.45484432 0.30665457 0.35786611 0.32363238
#>  [9,] 0.27424788 0.05130944 0.29105516 0.27381805 0.22779532 0.20403078
#> [10,] 0.45585861 0.05046574 0.48260109 0.53974966 0.50798839 0.48805989
#> [11,] 0.25090904 0.11822193 0.12436494 0.07641520 0.16952713 0.21659379
#> [12,] 0.17425589 0.06611642 0.14376547 0.19221244 0.17386328 0.18856172
#> [13,] 0.06395552 0.04163847 0.06866191 0.02872249 0.07764657 0.19055060
#> [14,] 0.13638285 0.04393532 0.18471018 0.16173342 0.12912920 0.09449709
#> [15,] 0.18912067 0.04656660 0.27955475 0.22976354 0.21779595 0.30198527
#> [16,] 0.27699103 0.03921926 0.19149250 0.14694096 0.06179152 0.03671038
#> [17,] 0.83015824 0.02655898 0.81225569 0.70152265 0.61865177 0.48820190
#> [18,] 0.08864433 0.04004254 0.11452839 0.07237302 0.03529630 0.08855182
#> [19,] 0.33749503 0.01330394 0.35898763 0.31647459 0.23792200 0.15361386
#> [20,] 0.06676069 0.02862138 0.17702969 0.20000362 0.20480284 0.27298141
#> [21,] 0.15753273 0.01683261 0.18973508 0.23772398 0.21430791 0.16656550
#> [22,] 0.20289788 0.01931793 0.41249123 0.28304016 0.37373786 0.26901218
#> [23,] 0.13269611 0.04113625 0.36378616 0.44719670 0.41424160 0.28537185
#> [24,] 0.21282397 0.04150786 0.12340813 0.11914519 0.13930367 0.23472406
#> [25,] 0.11636020 0.03060043 0.31552827 0.43318443 0.34995014 0.18382400
#>              V44         V45          V46          V47        V48          V49
#>  [1,] 0.20398777  0.16847270  0.099260601  0.090281045 0.07951167  0.053951610
#>  [2,] 0.13572692  0.09436843  0.107815612  0.127570877 0.14015831  0.090956789
#>  [3,] 0.44040804  0.46347561  0.211649418  0.127076860 0.15507697  0.093060225
#>  [4,] 0.10957332  0.19101058  0.190283281  0.199250912 0.18276770  0.073520864
#>  [5,] 0.07435109  0.01962563  0.222753326  0.214791194 0.11652332  0.034634636
#>  [6,] 0.26687947  0.19137112  0.086745552  0.060671744 0.02887155  0.029777796
#>  [7,] 0.36835518  0.32694495  0.225249083  0.127226958 0.05810315  0.025301563
#>  [8,] 0.19801151  0.23757613  0.111814687  0.130428662 0.18098823  0.097129490
#>  [9,] 0.18356948  0.12445999  0.086090933  0.114890317 0.09262893  0.065403620
#> [10,] 0.52014993  0.50946671  0.443835423  0.331422525 0.22741128  0.116145781
#> [11,] 0.11279310  0.08239886  0.065265273  0.071103063 0.07342812  0.047067918
#> [12,] 0.14557476  0.14851871  0.028875427  0.034952535 0.05704906  0.042891991
#> [13,] 0.24082872  0.23937520  0.243272643  0.187081055 0.11279596  0.058495563
#> [14,] 0.12186546  0.13241778  0.109029031  0.112434882 0.07379662  0.037724195
#> [15,] 0.18526929  0.06918719  0.061267961  0.060614768 0.04927716  0.009433598
#> [16,] 0.01963523 -0.02599768 -0.000675867  0.012325999 0.03121182  0.008353544
#> [17,] 0.35774647  0.36736059  0.341491707  0.266237732 0.18256165  0.106805086
#> [18,] 0.14202129  0.13096067  0.042624306  0.052237337 0.04603369  0.020769421
#> [19,] 0.09222355  0.09502329  0.094236172  0.081089598 0.04222523  0.016828702
#> [20,] 0.36138563  0.33896883  0.213635182  0.090317434 0.01731250  0.014722197
#> [21,] 0.10933381  0.09707365  0.087781021  0.048706164 0.03888672  0.030527475
#> [22,] 0.29311092  0.05111950 -0.016006787 -0.006161615 0.04328360 -0.003622156
#> [23,] 0.20774870  0.24231434  0.248221449  0.215397444 0.19105911  0.122011566
#> [24,] 0.16743499  0.03705964  0.028156670  0.033907936 0.03995952  0.008402995
#> [25,] 0.16956791  0.15333291  0.069480320  0.107435802 0.12120101  0.072062487
#>               V5         V50         V51          V52         V53         V54
#>  [1,] 0.06632199 0.016020335 0.015011502  0.013335899 0.004145800 0.004085351
#>  [2,] 0.11630040 0.025786874 0.019385705  0.017524475 0.007504123 0.008873689
#>  [3,] 0.07829535 0.019236883 0.030520567  0.017422343 0.004894745 0.008074316
#>  [4,] 0.06988653 0.005353513 0.019236103  0.019602734 0.010550252 0.011618029
#>  [5,] 0.07345874 0.016285319 0.025926302  0.016208058 0.011104654 0.002293057
#>  [6,] 0.15973989 0.011801591 0.008596217  0.006378743 0.008466188 0.020317787
#>  [7,] 0.04275474 0.013391147 0.007890505  0.006540612 0.008953119 0.012598866
#>  [8,] 0.03317512 0.023901278 0.023378451  0.017001148 0.016744937 0.024149176
#>  [9,] 0.09883310 0.029458967 0.020755604  0.016771403 0.016439255 0.013517040
#> [10,] 0.07299762 0.038352116 0.032909592  0.024192799 0.010499001 0.011068302
#> [11,] 0.14584576 0.024326880 0.024021322  0.019869234 0.020135075 0.020768837
#> [12,] 0.15089289 0.015893533 0.016735778  0.021082385 0.018791701 0.011537512
#> [13,] 0.08425373 0.013845850 0.029676507  0.016671509 0.006200563 0.002031666
#> [14,] 0.06905627 0.021556431 0.010087646  0.009392558 0.008471477 0.006601587
#> [15,] 0.06761431 0.009653070 0.008907004  0.005327211 0.006799307 0.007913119
#> [16,] 0.06309451 0.008501130 0.004494623  0.005173829 0.007096719 0.003854770
#> [17,] 0.05842733 0.039035372 0.024943963  0.010653162 0.013000022 0.011512806
#> [18,] 0.06865333 0.007747692 0.016149732  0.009808799 0.013721780 0.007744045
#> [19,] 0.02985512 0.019401425 0.014611337  0.012427679 0.013286550 0.012164466
#> [20,] 0.03030493 0.015953633 0.010036371  0.002562832 0.006001320 0.010586625
#> [21,] 0.03885948 0.010729669 0.009738794  0.012266490 0.005695020 0.005690685
#> [22,] 0.06892600 0.001996733 0.007418127 -0.002418072 0.009606298 0.016677430
#> [23,] 0.04430724 0.029629107 0.016187806  0.014489274 0.010971241 0.010250693
#> [24,] 0.02630025 0.019082503 0.017895133  0.015597396 0.016210707 0.005106799
#> [25,] 0.09451509 0.007935495 0.027819137  0.012608968 0.010353983 0.008486294
#>               V55         V56         V57         V58           V59         V6
#>  [1,] 0.002964765 0.005107118 0.003005781 0.006292521  3.613243e-03 0.11717063
#>  [2,] 0.005348211 0.009819400 0.007831027 0.008008552  9.634424e-03 0.14925585
#>  [3,] 0.010476131 0.004307822 0.005788340 0.004795301  4.730658e-03 0.11523250
#>  [4,] 0.014677567 0.005221078 0.009726051 0.009779195  8.244075e-03 0.07432084
#>  [5,] 0.006153802 0.002434043 0.003281954 0.003447556 -8.072064e-05 0.10354364
#>  [6,] 0.014499155 0.016013047 0.013198657 0.011663880  1.674588e-02 0.09860806
#>  [7,] 0.006801824 0.007453547 0.004546157 0.007127441  1.064540e-02 0.08107729
#>  [8,] 0.016813540 0.009912960 0.006754237 0.016206854  1.028009e-02 0.06234270
#>  [9,] 0.010355092 0.009828683 0.010833436 0.009507345  9.613200e-03 0.14618626
#> [10,] 0.008787974 0.008116661 0.009810697 0.006545608  7.156887e-03 0.09626216
#> [11,] 0.017127408 0.011720900 0.006786835 0.012449879  1.493972e-02 0.12600675
#> [12,] 0.002988345 0.010812996 0.005459894 0.004154693  3.712655e-03 0.17349121
#> [13,] 0.001092966 0.004748833 0.002910536 0.005484521  1.831846e-03 0.07973906
#> [14,] 0.007825272 0.009551555 0.007032939 0.003136724  4.552648e-03 0.09508290
#> [15,] 0.005166579 0.007805452 0.014792147 0.011226693  5.788092e-03 0.09118950
#> [16,] 0.003038845 0.003436575 0.005712290 0.005088026  2.275344e-03 0.08156137
#> [17,] 0.013515027 0.012287171 0.004584597 0.012708950  1.388318e-02 0.08426437
#> [18,] 0.002292221 0.007544045 0.004578755 0.003151288  4.328488e-03 0.08969513
#> [19,] 0.008220200 0.005238645 0.006172993 0.004455080  6.512074e-03 0.06790703
#> [20,] 0.001205497 0.001477372 0.003704182 0.003337717  6.183551e-03 0.02033388
#> [21,] 0.006082010 0.003547476 0.003262402 0.004581551  3.072119e-03 0.06795204
#> [22,] 0.010940641 0.017174427 0.017046082 0.006309956  9.976533e-03 0.10536538
#> [23,] 0.006783734 0.005034547 0.020904210 0.009501052  5.889262e-03 0.06988093
#> [24,] 0.004759412 0.007350036 0.003711157 0.004239044  1.772046e-02 0.05830211
#> [25,] 0.004577824 0.007129782 0.009510516 0.014525249  2.072399e-02 0.08856098
#>                 V60         V7         V8         V9
#>  [1,]  0.0008723953 0.11191182 0.10242596 0.18727072
#>  [2,]  0.0102418215 0.15832063 0.22291197 0.26892032
#>  [3,]  0.0044262496 0.11362317 0.17153246 0.25446277
#>  [4,]  0.0043727691 0.10882752 0.09213510 0.09725543
#>  [5,] -0.0009884831 0.11355076 0.11463030 0.15638862
#>  [6,]  0.0128084501 0.09396943 0.09141726 0.05939130
#>  [7,]  0.0054009195 0.12739375 0.16187204 0.32886540
#>  [8,]  0.0055860381 0.12229951 0.29759309 0.46629115
#>  [9,]  0.0065910422 0.18143221 0.18405658 0.21475880
#> [10,]  0.0042537194 0.10962777 0.14368504 0.23092653
#> [11,]  0.0102396273 0.13317547 0.11587087 0.14221186
#> [12,]  0.0023773738 0.13869131 0.17648214 0.19375714
#> [13,]  0.0031071140 0.08644136 0.12779434 0.21562852
#> [14,]  0.0049701425 0.12256696 0.10902383 0.12716547
#> [15,]  0.0063111746 0.08885263 0.10858790 0.06984904
#> [16,]  0.0026175555 0.05696621 0.09935080 0.13817034
#> [17,]  0.0117125271 0.13598983 0.15189600 0.03012039
#> [18,]  0.0029906658 0.13355032 0.10023277 0.13708085
#> [19,]  0.0055298486 0.06575601 0.11621218 0.13561446
#> [20,]  0.0019355896 0.08392520 0.07737771 0.07015528
#> [21,]  0.0028617126 0.08277539 0.08831073 0.10671841
#> [22,]  0.0102602144 0.12215811 0.12011255 0.09646749
#> [23,]  0.0048071074 0.19124128 0.28315154 0.28820540
#> [24,]  0.0093511552 0.04417823 0.10492919 0.16871631
#> [25,]  0.0154008077 0.16347637 0.16492842 0.13425797
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841