Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.026209371 0.46347975 0.43812173 0.48573378 0.48647603 0.51085963
#>  [2,] 0.035074459 0.05515093 0.13633728 0.25883912 0.30442065 0.34987246
#>  [3,] 0.057692082 0.19459440 0.24971097 0.24618366 0.30923196 0.37583110
#>  [4,] 0.017303455 0.08126989 0.16929837 0.22328328 0.14336570 0.13415205
#>  [5,] 0.027122351 0.30527280 0.36482661 0.40291053 0.37031706 0.30449913
#>  [6,] 0.079404345 0.54805815 0.39766081 0.18340429 0.19846617 0.34738388
#>  [7,] 0.027685781 0.27328363 0.33748448 0.34810584 0.31676500 0.22147622
#>  [8,] 0.061442000 0.25779429 0.35574658 0.48852150 0.55160640 0.67662772
#>  [9,] 0.015740416 0.30361190 0.28702667 0.31535112 0.24069982 0.19006338
#> [10,] 0.028523370 0.28945518 0.34976316 0.39030486 0.37865328 0.36716166
#> [11,] 0.018965521 0.18272166 0.15992121 0.10806252 0.12979864 0.17972935
#> [12,] 0.011437474 0.10566964 0.15155368 0.16017650 0.15851662 0.11581769
#> [13,] 0.028965451 0.08394577 0.13345640 0.23061338 0.25009714 0.16734330
#> [14,] 0.006785476 0.10685937 0.09676116 0.08867213 0.09765148 0.09950585
#> [15,] 0.008015891 0.18885438 0.14421702 0.08266599 0.20558963 0.28111138
#> [16,] 0.035300000 0.08800000 0.10780000 0.09790000 0.22500000 0.28190000
#> [17,] 0.026710067 0.11240561 0.09337439 0.07672762 0.04138637 0.11616690
#> [18,] 0.030633882 0.29014887 0.27027783 0.07433134 0.23127619 0.49636120
#> [19,] 0.019114385 0.13962512 0.15828231 0.21379970 0.22955143 0.21772682
#> [20,] 0.022875737 0.36995863 0.28115476 0.13602823 0.24913188 0.52331288
#> [21,] 0.016842151 0.13399545 0.13681312 0.09560908 0.09711184 0.12748058
#> [22,] 0.085600000 0.27510000 0.27070000 0.09460000 0.10200000 0.45190000
#> [23,] 0.022391613 0.15405566 0.18336927 0.21063921 0.31178224 0.36274262
#> [24,] 0.028940154 0.08306930 0.07880878 0.22993878 0.18269580 0.11942331
#> [25,] 0.019711717 0.10619590 0.12761711 0.11204383 0.12808466 0.11800385
#>              V15       V16       V17       V18       V19         V2       V20
#>  [1,] 0.44622388 0.5406868 0.7481378 0.7821744 0.7914358 0.12660432 0.8333371
#>  [2,] 0.40357498 0.4393495 0.5880549 0.7369596 0.8614851 0.05978985 0.9042862
#>  [3,] 0.32551198 0.3366857 0.3932178 0.2713027 0.3159586 0.06987740 0.5017359
#>  [4,] 0.19902815 0.3310492 0.3599400 0.2973789 0.4038460 0.04542643 0.3577028
#>  [5,] 0.28131001 0.3268064 0.3373411 0.4118740 0.5704608 0.02770218 0.6870461
#>  [6,] 0.42813770 0.4789755 0.2251240 0.2061801 0.2969369 0.08333424 0.8729007
#>  [7,] 0.18314702 0.1660451 0.1820431 0.2262520 0.2315621 0.02941673 0.2546476
#>  [8,] 0.77803192 0.8497230 1.0064538 0.9257798 0.9089586 0.08547126 0.7956187
#>  [9,] 0.12922574 0.1844233 0.2166017 0.2663721 0.2020062 0.03884196 0.1214826
#> [10,] 0.28419638 0.2291702 0.2515278 0.2851666 0.4029938 0.03749506 0.5895257
#> [11,] 0.34387674 0.4791210 0.5910736 0.5854027 0.6566384 0.02338395 0.8271105
#> [12,] 0.05101907 0.1175044 0.1978524 0.3234909 0.4133686 0.02640697 0.5466773
#> [13,] 0.10765901 0.1664105 0.2337301 0.3868842 0.3097262 0.01662073 0.3657638
#> [14,] 0.15407615 0.1649905 0.1278728 0.1479220 0.1325488 0.01251715 0.1513072
#> [15,] 0.35196151 0.2428840 0.1007039 0.1487600 0.2035179 0.01724579 0.2137217
#> [16,] 0.20990000 0.1240000 0.1699000 0.0939000 0.1091000 0.07130000 0.1410000
#> [17,] 0.08931306 0.1872059 0.1571209 0.2899011 0.3670587 0.03448143 0.3800433
#> [18,] 0.61975475 0.7375436 0.8409863 0.7268099 0.5079736 0.10390954 0.3803863
#> [19,] 0.25157718 0.3941141 0.4709657 0.3866060 0.3134318 0.01950096 0.3530172
#> [20,] 0.66464061 0.7557312 0.7951028 0.7132846 0.4872397 0.04667801 0.4004629
#> [21,] 0.10360336 0.1168633 0.2066265 0.2642254 0.2987997 0.01480691 0.4459274
#> [22,] 0.67370000 0.6699000 0.7066000 0.5632000 0.3785000 0.04540000 0.2721000
#> [23,] 0.48093639 0.5882115 0.5941372 0.5833542 0.6546357 0.03066063 0.7011651
#> [24,] 0.14908275 0.2118845 0.2178462 0.1880374 0.1869004 0.03516944 0.1861171
#> [25,] 0.08774094 0.1019426 0.1408333 0.2112723 0.3878130 0.02202608 0.3664105
#>              V21        V22       V23       V24       V25        V26       V27
#>  [1,] 0.75423318 0.83141043 1.0807724 1.1925884 1.0568967 1.01227871 1.0106067
#>  [2,] 0.88058845 0.91797320 0.9364994 0.9422431 0.7786095 0.62547526 0.5341291
#>  [3,] 0.59874044 0.66293885 0.6642650 0.6992857 0.6736285 0.65743723 0.7495763
#>  [4,] 0.47032435 0.44483509 0.4989781 0.5790736 0.4903121 0.53302175 0.6859281
#>  [5,] 0.78811580 0.84174254 0.8392039 0.8947616 0.9059994 0.94486748 0.9451204
#>  [6,] 0.98371647 0.39710713 0.1208981 0.1518836 0.2174668 0.40599563 0.2895048
#>  [7,] 0.32987124 0.45978947 0.5847230 0.6786940 0.7588475 0.84832013 0.9447855
#>  [8,] 0.80804779 0.72411292 0.4420249 0.1955918 0.1552421 0.07218804 0.1840874
#>  [9,] 0.06792342 0.17172078 0.3811944 0.5127722 0.6356378 0.80126539 0.9207184
#> [10,] 0.70994894 0.67352974 0.6281844 0.6625338 0.7035696 0.76657458 0.8847457
#> [11,] 0.94994102 0.99893179 0.8609656 0.6481599 0.4664225 0.34750811 0.2213745
#> [12,] 0.62528171 0.68588493 0.6781483 0.5912622 0.6637312 0.81924456 0.9335171
#> [13,] 0.63875978 0.69131032 0.7202072 0.8062578 0.7847724 0.68858252 0.5562877
#> [14,] 0.23178209 0.30202359 0.4420485 0.6364109 0.8298452 0.94764729 0.8170898
#> [15,] 0.21827454 0.37108202 0.5862681 0.7403948 0.6947107 0.72082457 0.7420345
#> [16,] 0.12680000 0.31510000 0.1430000 0.2264000 0.5756000 0.78760000 0.7158000
#> [17,] 0.31743001 0.15490421 0.1398543 0.2708590 0.1336398 0.24832558 0.4385004
#> [18,] 0.53752640 0.74043095 0.8242267 0.9627228 0.8355494 0.68742263 0.6726422
#> [19,] 0.38002578 0.48998933 0.5476691 0.5574198 0.6435434 0.75249838 0.7906952
#> [20,] 0.49159705 0.67453036 0.7927858 0.8033760 0.7061463 0.70899932 0.7185705
#> [21,] 0.58003482 0.75380606 0.8920690 0.8842793 0.8803715 0.92659042 0.8998779
#> [22,] 0.52970000 0.76970000 0.8643000 0.9304000 0.9372000 0.62470000 0.6024000
#> [23,] 0.71833957 0.68401170 0.6165152 0.6876553 0.7675165 0.79828002 0.7628983
#> [24,] 0.24044539 0.09631154 0.1896826 0.1969646 0.0864739 0.21573551 0.4708740
#> [25,] 0.24432655 0.26684702 0.3995571 0.5506106 0.5436563 0.50229079 0.4414201
#>             V28        V29          V3       V30         V31       V32
#>  [1,] 0.7257995 0.59057126 0.150673731 0.5160594 0.259177459 0.1751446
#>  [2,] 0.3802142 0.29223201 0.093727777 0.2508302 0.220146986 0.2182808
#>  [3,] 0.7545443 0.71445039 0.114249535 0.8011104 0.840628713 0.8589665
#>  [4,] 0.7068313 0.67800138 0.061520634 0.8036445 0.623114824 0.5166901
#>  [5,] 0.8812353 0.74549291 0.029506920 0.5686771 0.428760434 0.3481440
#>  [6,] 0.4694541 0.48355369 0.071025048 0.6758815 0.320383203 0.3660454
#>  [7,] 1.0369883 0.92888735 0.031909985 0.6828576 0.651319423 0.5645849
#>  [8,] 0.1974382 0.08278059 0.062422391 0.1061094 0.002489673 0.1081974
#>  [9,] 0.9383597 0.91753928 0.028070477 0.9606855 0.915249871 0.8350598
#> [10,] 0.8754613 0.90725256 0.052997977 0.8302082 0.679407641 0.6064274
#> [11,] 0.1027885 0.28781557 0.009899234 0.4096611 0.287837431 0.4149046
#> [12,] 0.9719400 0.81772882 0.016895095 0.5529179 0.282800530 0.1641678
#> [13,] 0.4116628 0.45183867 0.023622662 0.4239477 0.476937058 0.5590873
#> [14,] 0.6047187 0.49080480 0.007340763 0.3616170 0.274995554 0.1616942
#> [15,] 0.8722543 0.95904007 0.041815437 0.9016510 0.826130937 0.6969401
#> [16,] 0.5998000 0.55830000 0.032600000 0.6295000 0.765900000 0.8940000
#> [17,] 0.7273402 0.88043937 0.060913633 0.7745951 0.492916727 0.2424555
#> [18,] 0.5000839 0.40233533 0.125300607 0.4620293 0.379751445 0.1239747
#> [19,] 0.7868252 0.77503867 0.023313166 0.8008126 0.728354388 0.6181991
#> [20,] 0.5799988 0.38754057 0.074311193 0.4406597 0.548170143 0.4087293
#> [21,] 0.8422237 0.80816564 0.021967943 0.7055884 0.602327110 0.5463216
#> [22,] 0.6810000 0.50470000 0.038200000 0.5775000 0.475400000 0.2400000
#> [23,] 0.7160044 0.63171575 0.034368976 0.5194578 0.430776354 0.3120908
#> [24,] 0.6810050 0.71138453 0.032613252 0.6825641 0.739351321 0.6126496
#> [25,] 0.6482655 0.91498559 0.019794830 0.9648276 0.922801347 0.8541767
#>              V33        V34        V35        V36        V37       V38
#>  [1,] 0.04973639 0.28340342 0.44001066 0.27759918 0.24116904 0.3159989
#>  [2,] 0.33039466 0.34599129 0.33777259 0.33674490 0.43955873 0.4575035
#>  [3,] 0.61715120 0.52636911 0.56522345 0.32214825 0.20523657 0.3862579
#>  [4,] 0.40632603 0.40292807 0.26248788 0.10746605 0.27627970 0.4050836
#>  [5,] 0.28356347 0.21642834 0.13579543 0.14827469 0.20080421 0.2325493
#>  [6,] 0.52878229 0.69956452 0.75963355 0.78581779 0.76389220 0.8383951
#>  [7,] 0.46246881 0.31790730 0.17208533 0.07473040 0.15503820 0.1172569
#>  [8,] 0.24086867 0.15970679 0.32706926 0.03537873 0.15912323 0.2640653
#>  [9,] 0.73907307 0.67604820 0.70239995 0.74782097 0.72474495 0.5743000
#> [10,] 0.59126165 0.44109946 0.21426167 0.28916155 0.29543099 0.2651899
#> [11,] 0.40010479 0.20787695 0.05427647 0.22926751 0.37134202 0.3086121
#> [12,] 0.24850028 0.46228218 0.70668847 0.81667364 0.81166564 0.7043537
#> [13,] 0.70275078 0.63546068 0.60659591 0.59602268 0.55554180 0.4005667
#> [14,] 0.19475655 0.22523401 0.24365503 0.18556502 0.08637508 0.0851697
#> [15,] 0.63833982 0.55973652 0.46099009 0.43921397 0.25923921 0.1424272
#> [16,] 0.84360000 0.68070000 0.83800000 1.00000000 0.94970000 0.7866000
#> [17,] 0.52199133 0.58879200 0.33536900 0.42815145 0.59507559 0.6279981
#> [18,] 0.15966543 0.05319355 0.37156818 0.66766969 0.66040299 0.3274277
#> [19,] 0.62715560 0.68252163 0.76148388 0.74704323 0.61909806 0.4612208
#> [20,] 0.25063291 0.21988958 0.47279626 0.59136522 0.55635135 0.3469697
#> [21,] 0.55187678 0.59739392 0.61403357 0.54570151 0.44252752 0.3147515
#> [22,] 0.27790000 0.19970000 0.53050000 0.74090000 0.77750000 0.4424000
#> [23,] 0.28814377 0.29943661 0.28416983 0.25506305 0.24178702 0.2043095
#> [24,] 0.64974411 0.84728157 0.94472539 0.93304435 0.77151231 0.6773064
#> [25,] 0.77632090 0.70596473 0.62873272 0.54878429 0.49138562 0.2222507
#>              V39         V4        V40        V41       V42          V43
#>  [1,] 0.26694567 0.05561360 0.22479896 0.44873992 0.4116098 -0.007768937
#>  [2,] 0.38052020 0.06769900 0.39016993 0.37762036 0.4348804  0.326768968
#>  [3,] 0.41289376 0.11976658 0.21761699 0.53692084 0.5872563  0.354900128
#>  [4,] 0.31131507 0.06153491 0.42112342 0.41346500 0.2247597  0.276613393
#>  [5,] 0.28792750 0.04693513 0.27054237 0.16651653 0.2137428  0.281516414
#>  [6,] 0.70928972 0.10327264 0.52112902 0.37170706 0.3078957  0.285765054
#>  [7,] 0.16309068 0.04712828 0.20507810 0.26901790 0.2434286  0.188834391
#>  [8,] 0.19999205 0.13132266 0.01870685 0.03229356 0.2869574  0.247668641
#>  [9,] 0.46760321 0.05944384 0.49479678 0.46121508 0.4213912  0.474277705
#> [10,] 0.39533100 0.05428733 0.37156083 0.35076769 0.4618600  0.409886420
#> [11,] 0.21426684 0.03535054 0.12717435 0.13373651 0.1294839  0.065827879
#> [12,] 0.59350288 0.01534382 0.40558349 0.24846617 0.2084064  0.306419763
#> [13,] 0.40847338 0.01783805 0.63050676 0.68943864 0.6535550  0.636416478
#> [14,] 0.18513287 0.01477987 0.24165471 0.29034005 0.2877054  0.216135284
#> [15,] 0.16077224 0.03804361 0.10493518 0.12431581 0.1696450  0.261759921
#> [16,] 0.56470000 0.02720000 0.34800000 0.25850000 0.2304000  0.294800000
#> [17,] 0.57372667 0.06638659 0.37094773 0.23324116 0.1003886  0.038847282
#> [18,] 0.07141528 0.10904337 0.30344495 0.40325809 0.3668161  0.223396107
#> [19,] 0.33283888 0.03453535 0.37763554 0.34123829 0.2688057  0.217877953
#> [20,] 0.33837499 0.06369988 0.44267597 0.36141830 0.2887781  0.219053130
#> [21,] 0.18801941 0.03071448 0.21216393 0.20605564 0.1442031  0.155835639
#> [22,] 0.14160000 0.02030000 0.35080000 0.44820000 0.4208000  0.305400000
#> [23,] 0.21234885 0.03909953 0.20084781 0.18528237 0.1556290  0.133904979
#> [24,] 0.64452529 0.02556633 0.65798646 0.57484277 0.5073921  0.364523032
#> [25,] 0.08850118 0.03347973 0.14350267 0.17780493 0.1971342  0.243872046
#>                V44        V45         V46        V47        V48         V49
#>  [1,] 0.1435903037 0.20876374 0.205226453 0.09518830 0.05708673 0.006426627
#>  [2,] 0.1992794231 0.14033519 0.118428471 0.06237055 0.04676732 0.033254059
#>  [3,] 0.4648133628 0.59315595 0.521209019 0.31334860 0.16697507 0.082050929
#>  [4,] 0.3128942469 0.25899937 0.195249075 0.12256288 0.13082971 0.090882816
#>  [5,] 0.2903105398 0.27095531 0.199758684 0.13743466 0.09080215 0.054109781
#>  [6,] 0.2762829813 0.42639058 0.320002992 0.12174666 0.15007286 0.088604944
#>  [7,] 0.1114073418 0.11183012 0.134440911 0.15554286 0.14368105 0.070451311
#>  [8,] 0.0003442872 0.18439645 0.144920861 0.07147236 0.08612664 0.057739733
#>  [9,] 0.5309823899 0.47319993 0.319150706 0.21504354 0.17157727 0.103679287
#> [10,] 0.2798865626 0.29096164 0.137739606 0.11697221 0.14033657 0.076743169
#> [11,] 0.0834621866 0.07611421 0.032693230 0.04782256 0.05212539 0.016806350
#> [12,] 0.4003465116 0.38574804 0.307401247 0.20920248 0.12732525 0.069680666
#> [13,] 0.5253316975 0.32871538 0.211614342 0.14028153 0.10045965 0.027554019
#> [14,] 0.1433305789 0.11482785 0.119854091 0.06030182 0.03780672 0.030052604
#> [15,] 0.1746003919 0.03703470 0.032557498 0.03476480 0.03960960 0.014332803
#> [16,] 0.3363000000 0.30170000 0.219300000 0.13160000 0.10780000 0.055900000
#> [17,] 0.0195105476 0.01164549 0.008125774 0.02150940 0.02077069 0.006887354
#> [18,] 0.1969763053 0.21469285 0.152234826 0.15745419 0.16568804 0.102845898
#> [19,] 0.1898531077 0.18102656 0.178974013 0.10227570 0.05009562 0.022070331
#> [20,] 0.1729079267 0.17094310 0.135919725 0.12528748 0.10779216 0.061865598
#> [21,] 0.1008921976 0.06327053 0.055503311 0.03842304 0.03427552 0.012674977
#> [22,] 0.2235000000 0.26110000 0.279800000 0.23920000 0.20210000 0.132600000
#> [23,] 0.1078343786 0.12919902 0.123381016 0.09086784 0.05834617 0.028320282
#> [24,] 0.2200377082 0.14282881 0.074398331 0.09461248 0.09519086 0.052025135
#> [25,] 0.3518437978 0.32574307 0.213047388 0.10871234 0.03902809 0.023732295
#>               V5         V50         V51           V52           V53
#>  [1,] 0.15466836 0.018322180 0.018297729  0.0247788156  0.0281520843
#>  [2,] 0.13134276 0.012096029 0.011120944  0.0164497507  0.0183382802
#>  [3,] 0.09895468 0.039384818 0.033724572  0.0267120788  0.0186730463
#>  [4,] 0.02682464 0.021115849 0.019338597  0.0242274479  0.0134425634
#>  [5,] 0.07829804 0.016933646 0.019483238  0.0117949278  0.0100617499
#>  [6,] 0.08321208 0.038396113 0.024174397  0.0209787527  0.0182951960
#>  [7,] 0.07292664 0.013634609 0.019934847  0.0118477529  0.0071572974
#>  [8,] 0.23780584 0.054939965 0.031704091  0.0365683116  0.0289532464
#>  [9,] 0.08959216 0.029285896 0.012783052  0.0154891664  0.0103920382
#> [10,] 0.07831012 0.012813651 0.021292475  0.0164763187  0.0050585029
#> [11,] 0.04465080 0.008329619 0.013567623  0.0093556814 -0.0005697798
#> [12,] 0.03318990 0.021999157 0.012518634  0.0090240975  0.0049271285
#> [13,] 0.03622804 0.016494159 0.004610293 -0.0006876372  0.0081261441
#> [14,] 0.02614598 0.009452565 0.003108276  0.0075471028  0.0033720009
#> [15,] 0.02349005 0.016590356 0.018028591  0.0155996705  0.0153855528
#> [16,] 0.03700000 0.003500000 0.009800000  0.0163000000  0.0242000000
#> [17,] 0.04992733 0.003085187 0.001651227  0.0092283535  0.0240910800
#> [18,] 0.09374300 0.018264696 0.021423348  0.0106838605  0.0071203400
#> [19,] 0.05765996 0.022344667 0.016094511  0.0130181468  0.0081999899
#> [20,] 0.07071220 0.019015221 0.016790570  0.0128937673  0.0075322746
#> [21,] 0.05844904 0.008369717 0.005153102  0.0050191408  0.0089588453
#> [22,] 0.03850000 0.035800000 0.012800000  0.0172000000  0.0138000000
#> [23,] 0.07391131 0.018790398 0.011235904  0.0104530357  0.0092492236
#> [24,] 0.03987402 0.019836132 0.011770305  0.0062049022  0.0159767433
#> [25,] 0.03856609 0.017210470 0.012898357  0.0028972411  0.0060874763
#>               V54          V55         V56         V57         V58
#>  [1,] 0.038534721 0.0181587383 0.016483331 0.015509698 0.013053259
#>  [2,] 0.019434069 0.0073612958 0.013715727 0.009138026 0.007029863
#>  [3,] 0.025934933 0.0318390711 0.019727080 0.016503936 0.016080601
#>  [4,] 0.016616728 0.0184554106 0.015323556 0.013540408 0.011699477
#>  [5,] 0.008180663 0.0064439237 0.006748453 0.005016335 0.004015129
#>  [6,] 0.027436356 0.0170554364 0.023281944 0.013882770 0.016926884
#>  [7,] 0.006850266 0.0042416492 0.007808525 0.004144298 0.007771913
#>  [8,] 0.016863792 0.0231923696 0.007473396 0.004035384 0.013310800
#>  [9,] 0.008848803 0.0038795167 0.009004015 0.006625730 0.005507389
#> [10,] 0.011522801 0.0077545829 0.004345932 0.008208164 0.005122546
#> [11,] 0.005234781 0.0037190485 0.004183311 0.005805514 0.004658111
#> [12,] 0.004230271 0.0034830799 0.004501726 0.005198467 0.002882451
#> [13,] 0.007740551 0.0005812887 0.012818704 0.010990898 0.006818368
#> [14,] 0.003626649 0.0045938423 0.003011255 0.002476434 0.004002747
#> [15,] 0.004773123 0.0043795244 0.005688250 0.005874423 0.005117704
#> [16,] 0.004300000 0.0202000000 0.010800000 0.003700000 0.009600000
#> [17,] 0.019118127 0.0089759802 0.011614240 0.013227907 0.007383020
#> [18,] 0.015735274 0.0109177782 0.006382219 0.008712542 0.010135463
#> [19,] 0.011917599 0.0102014648 0.006504655 0.010618638 0.005879029
#> [20,] 0.011816163 0.0094317901 0.005296796 0.009459308 0.006695483
#> [21,] 0.006956959 0.0059585247 0.005854631 0.006912754 0.005713414
#> [22,] 0.007900000 0.0037000000 0.005100000 0.025800000 0.010200000
#> [23,] 0.011137634 0.0104938492 0.008486116 0.006528118 0.005239751
#> [24,] 0.013286289 0.0076477875 0.007567736 0.008174094 0.009972541
#> [25,] 0.011385798 0.0025776082 0.002759690 0.006150664 0.006050941
#>                 V59         V6         V60         V7         V8         V9
#>  [1,] -0.0008981553 0.21341460 0.002648304 0.18578818 0.28352605 0.43868684
#>  [2,]  0.0100880154 0.14797863 0.006936162 0.08245225 0.08923996 0.08568271
#>  [3,]  0.0275939116 0.10229711 0.017631934 0.08250340 0.07048643 0.21901494
#>  [4,]  0.0096311738 0.07011967 0.009290881 0.09882620 0.08926247 0.11620968
#>  [5,]  0.0056337200 0.11031329 0.005235773 0.12077374 0.14803073 0.22286265
#>  [6,]  0.0172864640 0.10920158 0.009711771 0.13473557 0.39712034 0.60527168
#>  [7,]  0.0068254884 0.11429735 0.004540909 0.11473202 0.12852535 0.18721766
#>  [8,]  0.0214515842 0.19707063 0.014339582 0.13788883 0.02397351 0.07327983
#>  [9,]  0.0026517953 0.10045006 0.003385747 0.14443518 0.17607482 0.29092366
#> [10,]  0.0048678324 0.10433559 0.005354778 0.07731215 0.10597313 0.21361483
#> [11,]  0.0016598281 0.07721674 0.001511692 0.11723236 0.11920963 0.19751138
#> [12,]  0.0039132949 0.04552178 0.003781827 0.05153813 0.01937685 0.06240259
#> [13,]  0.0097779889 0.08730488 0.008371547 0.10225893 0.09354668 0.06252807
#> [14,]  0.0011685511 0.04002544 0.003351794 0.06416072 0.06751422 0.10550691
#> [15,]  0.0151256096 0.05848065 0.009340734 0.03251660 0.09098433 0.15612428
#> [16,]  0.0093000000 0.07920000 0.005300000 0.10830000 0.06870000 0.02980000
#> [17,]  0.0128136666 0.09254199 0.010535107 0.01919296 0.06885569 0.13165957
#> [18,]  0.0126466814 0.10566441 0.008325402 0.18857702 0.20847292 0.24084360
#> [19,]  0.0093905712 0.09936769 0.007140935 0.12316453 0.09934619 0.08864255
#> [20,]  0.0089344640 0.13858499 0.008327398 0.11431397 0.19093941 0.27931167
#> [21,]  0.0051099796 0.09079507 0.005609996 0.09599416 0.08713364 0.11749616
#> [22,]  0.0037000000 0.05340000 0.003700000 0.21400000 0.31100000 0.28370000
#> [23,]  0.0054674943 0.09897373 0.004878806 0.11460150 0.10164248 0.13180832
#> [24,]  0.0123779140 0.06772769 0.007759370 0.05968538 0.07538706 0.07040547
#> [25,]  0.0071271876 0.04096969 0.001345438 0.09288492 0.08136396 0.08190139
#> 
#> $cl
#>  [1] M M M M M M M M M M M M R R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768