Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13        V14
#>  [1,] 0.02652352 0.19887686 0.24961151 0.28197905 0.32044159 0.39182206
#>  [2,] 0.04761926 0.06566089 0.15224732 0.25684784 0.31207115 0.39108756
#>  [3,] 0.03751529 0.27377216 0.40949776 0.48704333 0.51009781 0.47561727
#>  [4,] 0.01310000 0.31930000 0.34680000 0.37380000 0.30550000 0.19260000
#>  [5,] 0.05188956 0.23698771 0.30047622 0.33517899 0.36406421 0.35751630
#>  [6,] 0.01532474 0.22827887 0.28162903 0.26678789 0.25527285 0.19568459
#>  [7,] 0.05614770 0.06077456 0.15868280 0.22908088 0.26346750 0.32519040
#>  [8,] 0.01775120 0.22577383 0.26556087 0.32186003 0.30127043 0.24360951
#>  [9,] 0.03319054 0.47324611 0.55056919 0.53395130 0.48297410 0.36297527
#> [10,] 0.01861600 0.29107837 0.25193300 0.29518237 0.27301079 0.18569666
#> [11,] 0.02293195 0.19321615 0.20117881 0.26390296 0.29462360 0.26431660
#> [12,] 0.03696214 0.26780911 0.31039211 0.32733083 0.31434886 0.33974257
#> [13,] 0.10635402 0.56371236 0.46024945 0.19827960 0.19144089 0.27109463
#> [14,] 0.03949092 0.16931916 0.26495875 0.28884568 0.29002520 0.29008089
#> [15,] 0.01208999 0.08596162 0.10904693 0.13729756 0.22684100 0.23307126
#> [16,] 0.02944413 0.38963144 0.29612547 0.11500393 0.28982942 0.55276679
#> [17,] 0.01534156 0.10170576 0.10779931 0.10631657 0.11911146 0.14887858
#> [18,] 0.01894562 0.08726129 0.18394718 0.34345341 0.47283388 0.54712140
#> [19,] 0.02930000 0.08600000 0.04140000 0.04720000 0.08350000 0.09380000
#> [20,] 0.01711810 0.16041796 0.19669549 0.21149616 0.28070197 0.42088108
#> [21,] 0.02436243 0.08433286 0.08334844 0.10030808 0.19044966 0.33673177
#> [22,] 0.01366118 0.05894403 0.07693828 0.14395158 0.08861418 0.07475741
#> [23,] 0.02784621 0.11435220 0.07411225 0.21040277 0.19842164 0.10821322
#> [24,] 0.01646840 0.10288677 0.14897968 0.18962010 0.18752393 0.18132899
#> [25,] 0.02433297 0.11126659 0.15158593 0.06597294 0.11128847 0.14720220
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.45854879 0.5179953 0.5949961 0.6652949 0.7244933 0.021421788 0.7692753
#>  [2,] 0.47984134 0.5262557 0.6478554 0.7522428 0.8512941 0.050861648 0.8563343
#>  [3,] 0.55340981 0.6981929 0.8443777 0.9517992 0.9809912 0.062475876 0.9469669
#>  [4,] 0.13850000 0.2122000 0.2758000 0.4576000 0.6487000 0.038700000 0.7154000
#>  [5,] 0.31446075 0.3864650 0.3697848 0.3230461 0.3699330 0.077396135 0.6002417
#>  [6,] 0.10641655 0.1048546 0.2525168 0.3995336 0.5652091 0.033969340 0.6521033
#>  [7,] 0.42807068 0.3961602 0.5567179 0.6964752 0.8775587 0.089352802 0.9657411
#>  [8,] 0.18158452 0.2151975 0.2406451 0.3509568 0.5725842 0.024580286 0.6910388
#>  [9,] 0.24624124 0.2536890 0.2938755 0.3381804 0.5199233 0.034921071 0.6748489
#> [10,] 0.11624888 0.1944129 0.2142463 0.2384872 0.2061605 0.020346530 0.1547445
#> [11,] 0.21015945 0.1988588 0.1348990 0.1907686 0.3239304 0.029619416 0.3784660
#> [12,] 0.28035313 0.2522057 0.2694721 0.3040432 0.4637855 0.032166358 0.5301240
#> [13,] 0.54542771 0.5669478 0.1985398 0.1283178 0.3553748 0.104975642 0.8623810
#> [14,] 0.25979421 0.2150887 0.2386520 0.2833777 0.3219982 0.060365728 0.3247290
#> [15,] 0.19714703 0.2126314 0.2752600 0.3289982 0.4414325 0.008949588 0.5067644
#> [16,] 0.69192620 0.8886773 0.9240980 0.8313052 0.5576962 0.064502398 0.4177966
#> [17,] 0.19715632 0.1929346 0.1650581 0.1643480 0.1034482 0.016907988 0.1337313
#> [18,] 0.65759716 0.7192969 0.4884045 0.3521346 0.2649245 0.038325968 0.3501358
#> [19,] 0.14660000 0.0809000 0.1179000 0.2179000 0.3326000 0.064400000 0.3258000
#> [20,] 0.56833757 0.7213567 0.8599807 0.8657681 0.8499951 0.019016214 0.8150808
#> [21,] 0.49916423 0.6630017 0.7434774 0.7714166 0.8006021 0.021814949 0.8364285
#> [22,] 0.06150855 0.1481208 0.1383813 0.1485925 0.2218385 0.025669581 0.2492813
#> [23,] 0.17749646 0.2371040 0.2048269 0.1255609 0.1327389 0.037888943 0.1743902
#> [24,] 0.13302567 0.2028787 0.2809203 0.3312765 0.3635760 0.023296604 0.4585611
#> [25,] 0.19959557 0.1888765 0.1425638 0.1337023 0.1008012 0.022071805 0.1808043
#>              V21        V22       V23       V24        V25       V26        V27
#>  [1,] 0.84053449 0.90237643 0.9366165 0.9716946 0.94104616 0.9373359 0.92776406
#>  [2,] 0.93003775 1.04879085 1.0529694 0.9491244 0.74403946 0.6108580 0.59448611
#>  [3,] 0.84147820 0.58161833 0.3876188 0.3483872 0.29971994 0.1384918 0.09727713
#>  [4,] 0.80100000 0.79240000 0.8793000 1.0000000 0.98650000 0.9474000 0.94740000
#>  [5,] 0.64779104 0.64629554 0.6476657 0.6463633 0.64381030 0.7753426 0.88321575
#>  [6,] 0.65317629 0.54620221 0.4329109 0.4308691 0.58938620 0.8482071 1.00569368
#>  [7,] 0.95397801 1.03626045 0.9725972 0.9212296 0.85165216 0.7175233 0.66739821
#>  [8,] 0.79872431 0.79554746 0.8101549 0.9245517 0.94016897 0.9644568 0.98802013
#>  [9,] 0.79257145 0.86528592 0.8363425 0.8574970 0.82726585 0.8199221 0.85820577
#> [10,] 0.09352973 0.21442476 0.3969340 0.5434731 0.66164045 0.7592951 0.86362805
#> [11,] 0.42947160 0.58504611 0.6812151 0.7388773 0.82245430 0.8611563 0.93008621
#> [12,] 0.70344237 0.75124579 0.7621702 0.7905923 0.78986377 0.6764242 0.67742260
#> [13,] 0.88822306 0.40785003 0.1341047 0.1750536 0.21725947 0.3471424 0.25126864
#> [14,] 0.35854334 0.39381171 0.5225739 0.6510231 0.65764396 0.6976649 0.79647929
#> [15,] 0.52793430 0.56878320 0.6335322 0.6843691 0.78196777 0.8469385 0.92555278
#> [16,] 0.50407784 0.58028831 0.7485138 0.8609150 0.82349048 0.7292158 0.75029835
#> [17,] 0.21443359 0.21674890 0.3401353 0.5819251 0.84835392 0.9903845 0.84517530
#> [18,] 0.49330173 0.61135273 0.6412338 0.8596747 0.93998092 0.7743772 0.57970518
#> [19,] 0.21110000 0.23020000 0.3361000 0.4259000 0.46090000 0.2606000 0.08740000
#> [20,] 0.80826648 0.84989098 0.7929430 0.7517391 0.79389518 0.8700704 0.85226630
#> [21,] 0.97085444 0.95998362 0.9235953 0.9414753 0.91648813 0.9091925 0.81518268
#> [22,] 0.19875356 0.07517128 0.1825756 0.1386257 0.01811529 0.1980278 0.39792334
#> [23,] 0.23594237 0.16460002 0.2789309 0.2463847 0.10106768 0.2838165 0.62973213
#> [24,] 0.50885020 0.55125580 0.6457946 0.7323010 0.79368888 0.8021480 0.75347007
#> [25,] 0.30896179 0.54674144 0.8213680 0.8644447 0.82278868 0.8388766 0.80080553
#>             V28        V29         V3       V30       V31        V32       V33
#>  [1,] 0.8859607 0.76175315 0.03212579 0.6033663 0.4409227 0.32664395 0.2285094
#>  [2,] 0.4829314 0.45472242 0.06710387 0.4027207 0.2942434 0.19401998 0.2682729
#>  [3,] 0.0805092 0.06319544 0.08614437 0.1918986 0.2357752 0.30589377 0.3286086
#>  [4,] 0.9315000 0.83260000 0.03290000 0.6213000 0.3772000 0.28220000 0.2042000
#>  [5,] 0.8534751 0.69026869 0.08977752 0.7030088 0.6838134 0.69355904 0.4840819
#>  [6,] 0.9567131 0.85942756 0.03828112 0.7008665 0.4405883 0.38904251 0.5373174
#>  [7,] 0.4591095 0.25433042 0.21596619 0.2094382 0.1329853 0.06057398 0.1040994
#>  [8,] 0.9832697 0.87525621 0.03536445 0.6428670 0.4106058 0.29796570 0.2341574
#>  [9,] 0.8107015 0.62287130 0.04302679 0.4336872 0.3656515 0.37709854 0.4145597
#> [10,] 0.9017594 0.88765677 0.02477104 0.9439309 0.9472023 0.86959090 0.7355629
#> [11,] 0.9783085 0.90105543 0.03529818 0.7113552 0.6437360 0.53523743 0.4587243
#> [12,] 0.6957820 0.85016305 0.03554707 0.9503096 0.8453325 0.65523579 0.6220462
#> [13,] 0.5065888 0.51156375 0.05153122 0.6676837 0.3598381 0.41092898 0.6383358
#> [14,] 0.8596086 0.80786568 0.07478267 0.6167567 0.5500150 0.43421348 0.3536257
#> [15,] 0.8889896 0.71461908 0.01076750 0.5301817 0.4468210 0.36367368 0.2578907
#> [16,] 0.6088812 0.30799133 0.06943175 0.3486593 0.5714361 0.51503375 0.2706457
#> [17,] 0.6386403 0.52723934 0.01739509 0.4105940 0.3191401 0.17162282 0.1672458
#> [18,] 0.6474062 0.85428899 0.03238934 0.6873575 0.5265297 0.49938172 0.3015809
#> [19,] 0.2862000 0.56060000 0.03900000 0.8344000 0.8096000 0.72500000 0.8048000
#> [20,] 0.7792865 0.72560468 0.01809364 0.6180853 0.4175545 0.29914115 0.2327512
#> [21,] 0.5948416 0.39118013 0.03076648 0.4252458 0.4509083 0.35368566 0.3192808
#> [22,] 0.6453168 0.73422834 0.03941174 0.6955567 0.6544139 0.46393495 0.5965825
#> [23,] 0.8205561 0.82760114 0.01377760 0.7999536 0.7742553 0.57517779 0.6633510
#> [24,] 0.7506080 0.76911307 0.03028516 0.7089701 0.6839969 0.66189792 0.6287738
#> [25,] 0.8513201 0.94377250 0.02887848 0.8052269 0.6536688 0.52990778 0.5797112
#>             V34       V35        V36        V37        V38        V39
#>  [1,] 0.1047911 0.1308123 0.14233222 0.17799685 0.22087250 0.19246025
#>  [2,] 0.3084676 0.3573667 0.27015835 0.27112487 0.30661734 0.29237342
#>  [3,] 0.4266827 0.4666197 0.23311895 0.28654647 0.36921440 0.20555430
#>  [4,] 0.2190000 0.2223000 0.13270000 0.05210000 0.06180000 0.14160000
#>  [5,] 0.3769054 0.2880805 0.26914437 0.32769446 0.40813584 0.49055604
#>  [6,] 0.5785877 0.6129984 0.57589904 0.57370330 0.51400335 0.40625880
#>  [7,] 0.2213165 0.4527840 0.48735233 0.40145087 0.43689672 0.40508778
#>  [8,] 0.1805189 0.1552326 0.07426984 0.07546315 0.10384163 0.09426632
#>  [9,] 0.3847548 0.2208052 0.29541811 0.33058495 0.31840220 0.43905738
#> [10,] 0.6493311 0.6540061 0.69003795 0.68971751 0.59062093 0.54729826
#> [11,] 0.3710645 0.1871106 0.11865184 0.16019814 0.16975285 0.26363795
#> [12,] 0.5775737 0.3708147 0.31769788 0.18123545 0.08603609 0.25671618
#> [13,] 0.7475932 0.8817830 0.85694435 0.85700425 0.88758677 0.74215014
#> [14,] 0.2598835 0.1539492 0.18420803 0.25178204 0.18828787 0.22711402
#> [15,] 0.1675966 0.1514481 0.23262240 0.23488141 0.27328066 0.24529418
#> [16,] 0.3385338 0.7099810 0.70466279 0.56000831 0.52578004 0.37592610
#> [17,] 0.2044465 0.2724089 0.20829551 0.07992563 0.06100839 0.11525592
#> [18,] 0.3025356 0.3237885 0.26457402 0.08861101 0.08809693 0.13341808
#> [19,] 0.9435000 1.0000000 0.89600000 0.55160000 0.30370000 0.23380000
#> [20,] 0.2865052 0.2522616 0.20704010 0.22240571 0.17230818 0.08987461
#> [21,] 0.2811584 0.1627188 0.07250957 0.14625178 0.13518264 0.12225862
#> [22,] 0.7299326 0.6929531 0.73295916 0.71028445 0.69477998 0.69076343
#> [23,] 0.8870432 0.9604202 0.90756001 0.67657276 0.55633805 0.49810984
#> [24,] 0.6300235 0.6480667 0.61750637 0.54305082 0.35561251 0.23882046
#> [25,] 0.7117373 0.6883144 0.54107984 0.29328077 0.17867165 0.18240448
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.05138364 0.22461132 0.24292659 0.21398366 0.19404069 0.17153924
#>  [2,] 0.04704672 0.31650202 0.27191118 0.37015214 0.34181229 0.22546869
#>  [3,] 0.08797287 0.20937769 0.35576099 0.29884075 0.16570081 0.10231469
#>  [4,] 0.00780000 0.14600000 0.08460000 0.10550000 0.16390000 0.19160000
#>  [5,] 0.07771443 0.34386396 0.48512915 0.62113161 0.52189842 0.40544060
#>  [6,] 0.05658302 0.29401104 0.32417382 0.34539972 0.36073013 0.30444102
#>  [7,] 0.28096643 0.28348589 0.31569245 0.49161214 0.37212055 0.13160592
#>  [8,] 0.03894541 0.09578928 0.05597762 0.10840718 0.20216460 0.23062838
#>  [9,] 0.03990681 0.39605639 0.17902596 0.23836549 0.35828179 0.39479472
#> [10,] 0.03886055 0.56811001 0.52382014 0.42163539 0.45546383 0.53657695
#> [11,] 0.01415518 0.28960915 0.24958215 0.17919417 0.13038679 0.05785596
#> [12,] 0.05420626 0.36408135 0.40137542 0.37125116 0.28778086 0.06302580
#> [13,] 0.07807603 0.50718111 0.25706524 0.21184371 0.21113959 0.29246319
#> [14,] 0.10434888 0.27299572 0.27438723 0.35174004 0.24239164 0.12320741
#> [15,] 0.01400901 0.17312499 0.13831855 0.19771389 0.18653480 0.15491364
#> [16,] 0.05633412 0.30168739 0.17985630 0.21082941 0.26879466 0.25632892
#> [17,] 0.01747047 0.18730180 0.26312523 0.24460065 0.18653553 0.10381242
#> [18,] 0.03694154 0.11348956 0.17599475 0.10111903 0.16531616 0.18015145
#> [19,] 0.01730000 0.23820000 0.33180000 0.38210000 0.15750000 0.22280000
#> [20,] 0.03293637 0.12214538 0.10236866 0.08192270 0.08575906 0.06982652
#> [21,] 0.04184959 0.17672151 0.12215647 0.06629369 0.05029246 0.12010263
#> [22,] 0.01368405 0.63506888 0.47277381 0.41037369 0.27743526 0.14500852
#> [23,] 0.01212208 0.65952754 0.59759230 0.54113116 0.36366121 0.19369502
#> [24,] 0.03597588 0.28774597 0.25585662 0.22676644 0.18703063 0.18661772
#> [25,] 0.04813029 0.23904658 0.20485291 0.22451466 0.31065723 0.18378411
#>              V45        V46        V47        V48        V49         V5
#>  [1,] 0.13238232 0.09089138 0.09501233 0.07317895 0.04850203 0.10592007
#>  [2,] 0.14641908 0.12177340 0.12789119 0.05907757 0.03881414 0.08749936
#>  [3,] 0.05282491 0.04406348 0.06839424 0.04507452 0.02860620 0.07197233
#>  [4,] 0.20850000 0.23350000 0.19640000 0.13000000 0.06330000 0.07210000
#>  [5,] 0.55121994 0.44476877 0.25873704 0.18600647 0.12144390 0.12363689
#>  [6,] 0.22116114 0.15752697 0.10508642 0.06332621 0.03323053 0.11535785
#>  [7,] 0.18321294 0.13088727 0.04758638 0.08023508 0.04524457 0.34676052
#>  [8,] 0.20886084 0.20997756 0.16478588 0.10143239 0.05121273 0.06769561
#>  [9,] 0.35279988 0.17529427 0.10266094 0.10893443 0.06383512 0.06786166
#> [10,] 0.47875266 0.35545984 0.25667069 0.20120499 0.10482170 0.07817482
#> [11,] 0.04704783 0.06381850 0.08474948 0.08764787 0.04472933 0.03631950
#> [12,] 0.09758883 0.05522276 0.08857179 0.10636729 0.07094415 0.05715590
#> [13,] 0.46129017 0.33960876 0.15173444 0.09448167 0.05920186 0.07299750
#> [14,] 0.16004570 0.19185940 0.20391744 0.16200507 0.07433441 0.11436756
#> [15,] 0.11236795 0.03982668 0.06419315 0.07136295 0.02852077 0.03617183
#> [16,] 0.21745179 0.11953598 0.09804100 0.08379336 0.04926819 0.09324776
#> [17,] 0.09043552 0.08771104 0.04313080 0.03921431 0.03182151 0.03859962
#> [18,] 0.10745484 0.17339270 0.13194927 0.03602040 0.03163369 0.08093033
#> [19,] 0.15820000 0.14330000 0.16340000 0.11330000 0.05670000 0.04760000
#> [20,] 0.10371740 0.08915335 0.09310680 0.06599956 0.03619440 0.05745900
#> [21,] 0.10301675 0.05186997 0.07682633 0.05842620 0.02855926 0.07307024
#> [22,] 0.13176346 0.09298396 0.10964244 0.08820610 0.05521617 0.01482640
#> [23,] 0.16685149 0.10340880 0.11657679 0.08764732 0.04526888 0.03565358
#> [24,] 0.14343490 0.11136430 0.05838291 0.04249893 0.01980131 0.06901443
#> [25,] 0.05638876 0.05168498 0.06090349 0.05159057 0.01075483 0.06157482
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.020577219 0.016512385 0.016026992 0.014156483 0.012415597 0.005720186
#>  [2,] 0.017754261 0.011536644 0.012325605 0.017628224 0.018440857 0.000554994
#>  [3,] 0.019064972 0.016527412 0.029493294 0.021990054 0.016053225 0.026038869
#>  [4,] 0.018300000 0.013700000 0.015000000 0.007600000 0.003200000 0.003700000
#>  [5,] 0.047059328 0.028758209 0.019423903 0.013732503 0.011867206 0.018765622
#>  [6,] 0.011762886 0.009850199 0.010609898 0.009048018 0.005913421 0.001824022
#>  [7,] 0.010299460 0.030118600 0.024449115 0.020847072 0.026992158 0.007847296
#>  [8,] 0.014797623 0.021454111 0.014243391 0.005801458 0.003846184 0.003859770
#>  [9,] 0.016232619 0.019018144 0.012915586 0.009787849 0.011352466 0.009257964
#> [10,] 0.034348540 0.020636317 0.014713718 0.008193001 0.009178170 0.007584475
#> [11,] 0.007862466 0.008293516 0.012031412 0.009868545 0.004493085 0.005372243
#> [12,] 0.008979948 0.013873752 0.014927722 0.003579716 0.009027012 0.002190451
#> [13,] 0.030737489 0.018239302 0.018338345 0.011223735 0.021157738 0.022031014
#> [14,] 0.015483641 0.020382051 0.018918122 0.011350135 0.010524603 0.009201411
#> [15,] 0.010857197 0.005325906 0.010354815 0.006775750 0.005845011 0.007029920
#> [16,] 0.027652627 0.007248338 0.018406913 0.006629667 0.017285067 0.011102197
#> [17,] 0.010965623 0.007639399 0.010814428 0.004376629 0.004837630 0.006532781
#> [18,] 0.025049626 0.015439655 0.011445617 0.011396905 0.004973640 0.006893581
#> [19,] 0.013300000 0.017000000 0.003500000 0.005200000 0.008300000 0.007800000
#> [20,] 0.019104805 0.012413251 0.007070339 0.003362725 0.011561852 0.008615975
#> [21,] 0.014112667 0.011361714 0.010786078 0.010174527 0.005939599 0.004942743
#> [22,] 0.021622608 0.008010045 0.009722237 0.019690457 0.016758802 0.012398777
#> [23,] 0.024247681 0.015672651 0.010205432 0.018214851 0.015559280 0.007734616
#> [24,] 0.010118605 0.007700736 0.005705648 0.008009447 0.009614734 0.009432719
#> [25,] 0.010576590 0.011670594 0.006246775 0.006753393 0.006864075 0.004656531
#>               V56         V57         V58         V59         V6         V60
#>  [1,] 0.011652845 0.008130734 0.008719631 0.006714333 0.13839453 0.004509778
#>  [2,] 0.012683145 0.008540993 0.012712202 0.013629446 0.11006747 0.010920400
#>  [3,] 0.015417645 0.005680740 0.017509079 0.014406985 0.03659506 0.009276953
#>  [4,] 0.007100000 0.004000000 0.000900000 0.001500000 0.13410000 0.008500000
#>  [5,] 0.012831161 0.011481822 0.012481240 0.011261702 0.11992937 0.009090486
#>  [6,] 0.002314106 0.002587688 0.003459025 0.005774417 0.07081551 0.002567551
#>  [7,] 0.016143766 0.018188367 0.015622271 0.013754973 0.23844770 0.005433621
#>  [8,] 0.004075181 0.003308663 0.003795452 0.002283735 0.07159971 0.002940119
#>  [9,] 0.007238643 0.005072165 0.008116124 0.007931825 0.13114003 0.005596113
#> [10,] 0.008192703 0.007660943 0.003397106 0.003717871 0.09377131 0.003245195
#> [11,] 0.005812609 0.004357947 0.003971696 0.003794820 0.06211511 0.004754271
#> [12,] 0.004898333 0.008343609 0.006832197 0.005456777 0.12987593 0.007091758
#> [13,] 0.010445611 0.010953193 0.015537507 0.012529240 0.12481168 0.011757411
#> [14,] 0.006429984 0.007630437 0.009947275 0.008883727 0.12726220 0.005215118
#> [15,] 0.006089578 0.005287846 0.006703657 0.004386336 0.06000118 0.003986180
#> [16,] 0.003452330 0.008317436 0.005375967 0.009747708 0.20013553 0.006809309
#> [17,] 0.002653944 0.003477618 0.002937920 0.001688556 0.06807870 0.002246402
#> [18,] 0.012268882 0.008818280 0.005404985 0.004003095 0.10663684 0.002457250
#> [19,] 0.007500000 0.010500000 0.016000000 0.009500000 0.08160000 0.001100000
#> [20,] 0.007528443 0.006855644 0.004140384 0.005290266 0.09497664 0.006440463
#> [21,] 0.010847147 0.007541841 0.002527027 0.005260996 0.09511625 0.004433532
#> [22,] 0.006377439 0.007345623 0.011778799 0.014732386 0.04176081 0.001389100
#> [23,] 0.008445947 0.007556969 0.006405536 0.007182076 0.05758673 0.007226738
#> [24,] 0.008540418 0.007558155 0.004734349 0.006489868 0.10848306 0.006034764
#> [25,] 0.007122306 0.011916889 0.010139202 0.007484933 0.08930319 0.006455444
#>               V7         V8         V9
#>  [1,] 0.14204698 0.15186617 0.19058887
#>  [2,] 0.12386205 0.07673820 0.06879450
#>  [3,] 0.06869462 0.03481924 0.12201364
#>  [4,] 0.16260000 0.19020000 0.26100000
#>  [5,] 0.10834424 0.14894757 0.21786101
#>  [6,] 0.05369004 0.07811637 0.22381712
#>  [7,] 0.05744827 0.24679860 0.09047112
#>  [8,] 0.10311933 0.12254128 0.17968581
#>  [9,] 0.12946425 0.19838396 0.33794293
#> [10,] 0.13072935 0.14497728 0.24215113
#> [11,] 0.07486486 0.07551887 0.14432018
#> [12,] 0.08019496 0.09234286 0.19127983
#> [13,] 0.24379218 0.42250178 0.56247551
#> [14,] 0.13490762 0.12299293 0.13646640
#> [15,] 0.09149373 0.07745563 0.06490568
#> [16,] 0.22158567 0.22757325 0.32164021
#> [17,] 0.08027338 0.07704500 0.11101595
#> [18,] 0.11864108 0.11164962 0.10837311
#> [19,] 0.09930000 0.03150000 0.07360000
#> [20,] 0.12412614 0.09833441 0.13907402
#> [21,] 0.12114736 0.09656628 0.10772144
#> [22,] 0.03783622 0.08614311 0.04193955
#> [23,] 0.05775149 0.09295286 0.09140544
#> [24,] 0.11204221 0.09863267 0.09548758
#> [25,] 0.08140318 0.10997310 0.09118276
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058