Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1        V10        V11        V12         V13        V14
#>  [1,]  0.023365020 0.31814721 0.36760637 0.37878317  0.31374364 0.23044618
#>  [2,]  0.031879539 0.28095685 0.31876489 0.35919700  0.35033167 0.34896946
#>  [3,]  0.018595039 0.19704884 0.24768505 0.30376220  0.42577844 0.57199409
#>  [4,]  0.020740830 0.36775599 0.33636375 0.36224560  0.40517892 0.46888717
#>  [5,]  0.026725863 0.22383497 0.23100221 0.24891966  0.30140633 0.39803737
#>  [6,]  0.063619917 0.02433458 0.14664827 0.25534756  0.30706376 0.36145268
#>  [7,]  0.058694367 0.35463945 0.29181365 0.18551154  0.20163227 0.35290131
#>  [8,]  0.062039532 0.25977491 0.38359178 0.50579971  0.61648308 0.70490034
#>  [9,]  0.019062001 0.18125744 0.21013667 0.25474372  0.30757593 0.36909681
#> [10,]  0.023719859 0.22306393 0.28274830 0.32901887  0.30321457 0.31636676
#> [11,]  0.041834504 0.16908720 0.23196336 0.25781802  0.34389296 0.30033196
#> [12,]  0.013358508 0.40127438 0.44825606 0.38091535  0.28688816 0.16377497
#> [13,]  0.008876983 0.10750941 0.09645988 0.09444577  0.10732767 0.11240312
#> [14,]  0.011250608 0.12712273 0.11736213 0.02991134  0.06571461 0.05994707
#> [15,]  0.018053158 0.06710155 0.15259886 0.34320940  0.34745124 0.28064653
#> [16,]  0.025516388 0.12588034 0.15099641 0.16512851  0.24433938 0.38364232
#> [17,]  0.012472806 0.04386541 0.12471110 0.19502057  0.28373607 0.35645087
#> [18,]  0.028672345 0.09939591 0.22249257 0.33097696  0.26141820 0.21338116
#> [19,]  0.018989542 0.02717181 0.05938014 0.17734967  0.17741262 0.11756041
#> [20,]  0.020574045 0.06977681 0.07629062 0.07685450  0.11024298 0.14438112
#> [21,] -0.008598391 0.09229414 0.09111779 0.02902069 -0.01410696 0.05026441
#> [22,]  0.019467635 0.41385598 0.35611829 0.25722244  0.34268474 0.51774972
#> [23,]  0.008072721 0.15113964 0.11900128 0.12637021  0.21376568 0.21552303
#> [24,]  0.040008479 0.05414701 0.12896697 0.24121644  0.24559453 0.09044118
#> [25,]  0.018890425 0.13220554 0.13553703 0.10600050  0.09590734 0.13753275
#>              V15        V16        V17       V18       V19           V2
#>  [1,] 0.17606922 0.17418017 0.19859481 0.2952233 0.4690189  0.041208317
#>  [2,] 0.25594396 0.30829708 0.32702875 0.3326491 0.4209139  0.039848041
#>  [3,] 0.61654620 0.60470563 0.67102369 0.7556840 0.8339249  0.039441191
#>  [4,] 0.48819950 0.63758266 0.83215487 0.8363806 0.8341946  0.093738462
#>  [5,] 0.48825526 0.53420752 0.66608452 0.7142972 0.7195294  0.012095738
#>  [6,] 0.47755192 0.48416674 0.68755470 0.8484652 0.9607983  0.066252933
#>  [7,] 0.45335623 0.46671437 0.41029652 0.3399405 0.3770878  0.067906097
#>  [8,] 0.78277161 0.83334894 0.96179276 0.9404234 0.9086588  0.062917616
#>  [9,] 0.42834990 0.52759058 0.60053041 0.6707934 0.7108944  0.010757516
#> [10,] 0.33716990 0.43581798 0.51457621 0.6122053 0.6933176  0.012291113
#> [11,] 0.22711553 0.27764306 0.31873515 0.1940159 0.2440483  0.070793876
#> [12,] 0.14783544 0.24631597 0.34619371 0.4494748 0.5553933  0.017553985
#> [13,] 0.15871999 0.17279303 0.13683964 0.1475958 0.1180785  0.012940171
#> [14,] 0.10637781 0.06600567 0.10793420 0.2215702 0.3539392  0.014260479
#> [15,] 0.08770031 0.17600965 0.27229106 0.4294730 0.4186563  0.025788787
#> [16,] 0.55428563 0.72904807 0.83292497 0.8508339 0.8611445  0.027517085
#> [17,] 0.47043143 0.54097484 0.53823014 0.5689414 0.5952922  0.013580550
#> [18,] 0.24605628 0.37357436 0.52829404 0.4507144 0.4814846  0.054615456
#> [19,] 0.14349084 0.21454808 0.19686094 0.2294651 0.3080061  0.030251290
#> [20,] 0.11045013 0.12495503 0.15562237 0.2188381 0.3136087  0.029470147
#> [21,] 0.12807562 0.08127182 0.03178559 0.1433463 0.2598540 -0.017251194
#> [22,] 0.64486490 0.80630385 0.86722555 0.8165228 0.5953212  0.036852516
#> [23,] 0.29972967 0.33046618 0.29756143 0.2731522 0.1393693  0.016020564
#> [24,] 0.05340601 0.16176855 0.20855849 0.3346751 0.1188654  0.029170494
#> [25,] 0.11907591 0.19001079 0.34035526 0.3422482 0.3220430  0.008558429
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.5998577 0.7469177 0.7802547 0.7913298 0.9120637 0.9669242 1.0269968
#>  [2,] 0.6079856 0.6458288 0.6067577 0.5443960 0.5764992 0.5957506 0.8132211
#>  [3,] 0.9036965 0.9753714 0.9798226 0.9015082 0.8842542 0.8771759 0.8477942
#>  [4,] 0.8613288 0.8119931 0.8652614 1.0227803 1.0711058 0.9819474 0.9871874
#>  [5,] 0.7554563 0.8452480 0.8752768 0.9208497 0.9965977 1.0237374 0.9945136
#>  [6,] 0.9659342 0.9535349 0.9933447 0.9955724 0.9333732 0.7606559 0.5969090
#>  [7,] 0.7481367 0.9855561 0.8345220 0.5551334 0.3559457 0.2656064 0.4951283
#>  [8,] 0.7852097 0.7226694 0.6740286 0.5104332 0.3012235 0.2571046 0.2606808
#>  [9,] 0.7956735 0.8886259 0.9686007 1.0112967 1.0057174 0.9647909 0.9321292
#> [10,] 0.7846709 0.8433987 0.9046020 0.9202189 0.9136070 0.9502079 0.9211431
#> [11,] 0.2603357 0.3381472 0.4699349 0.5667813 0.5848539 0.5695725 0.6009906
#> [12,] 0.6233077 0.7109628 0.8196806 0.9100035 0.9829168 0.9487498 0.8308537
#> [13,] 0.1345253 0.2202786 0.2728068 0.4094969 0.6150223 0.8363421 0.9741377
#> [14,] 0.3800171 0.4633996 0.8146291 0.9066095 0.8533927 0.7100436 0.7239562
#> [15,] 0.6160332 0.6697891 0.6861938 0.7509726 0.7405961 0.9015715 0.8743726
#> [16,] 0.8533128 0.9177668 0.9156083 0.8468584 0.8386015 0.8492150 0.8771551
#> [17,] 0.6750864 0.8293420 0.9111764 0.9064539 0.9411496 0.9045997 0.7888345
#> [18,] 0.5919110 0.4172034 0.2973147 0.4178684 0.8760710 0.9065574 0.5025539
#> [19,] 0.1255591 0.2616619 0.2483387 0.2568976 0.3209349 0.2509651 0.3253700
#> [20,] 0.3221315 0.2710044 0.2987587 0.3216053 0.4228275 0.4321532 0.4531793
#> [21,] 0.3213452 0.3530765 0.5100494 0.7250916 0.9104896 0.9944385 0.8236819
#> [22,] 0.5164606 0.5619144 0.5905215 0.6791904 0.7525081 0.6906373 0.5916110
#> [23,] 0.1834947 0.3436658 0.2909351 0.2747883 0.5359586 0.8653196 0.9928510
#> [24,] 0.2737520 0.5844910 0.6823428 0.8351534 0.8645475 0.7783516 0.6281680
#> [25,] 0.4239599 0.5631869 0.7086610 0.8368546 0.8593380 0.8925330 0.9697966
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 1.0277641 0.9757923 0.8951675  0.03900869 0.6782553 0.4825028 0.3437570
#>  [2,] 0.9626396 0.8983178 0.7608959  0.06098491 0.7052493 0.5952002 0.6690052
#>  [3,] 0.7481650 0.6323543 0.5488411  0.03208335 0.3961017 0.2729539 0.1662124
#>  [4,] 0.9684406 0.7327816 0.6108373  0.10647142 0.5235755 0.3036511 0.2297009
#>  [5,] 0.9442364 0.8512876 0.6995265  0.02770931 0.5607456 0.4296943 0.3434885
#>  [6,] 0.5105174 0.4379011 0.4177212  0.11023006 0.3423581 0.2316038 0.1386780
#>  [7,] 0.5016335 0.3440240 0.2899458  0.05980197 0.4584544 0.2104787 0.2748089
#>  [8,] 0.2956849 0.3045094 0.1777493  0.05797552 0.2017391 0.1641103 0.2346377
#>  [9,] 0.9375784 0.9364588 0.7877950  0.01941017 0.5876561 0.3939005 0.3713514
#> [10,] 0.8525302 0.7160966 0.6177217  0.02007939 0.4878819 0.2694491 0.2361599
#> [11,] 0.6459972 0.8392765 0.8723858  0.05873175 0.8271304 0.9099551 0.8771908
#> [12,] 0.6171246 0.4105551 0.2456071  0.02785675 0.3627904 0.4532157 0.4524588
#> [13,] 0.8280985 0.5991254 0.4899630  0.00866794 0.3638633 0.2816523 0.1443503
#> [14,] 0.6143797 0.6886196 0.8114398  0.03606980 0.7913006 0.6291932 0.6606272
#> [15,] 0.6956543 0.5219209 0.4378062  0.06090963 0.3874046 0.4195286 0.4931513
#> [16,] 0.8235677 0.6633115 0.5176547  0.03379987 0.4932976 0.4426971 0.3284877
#> [17,] 0.6273429 0.4912838 0.5771670  0.01145498 0.6328517 0.4944567 0.3679848
#> [18,] 0.3850094 0.6898555 0.7806196  0.05752422 0.4023553 0.4632482 0.6247487
#> [19,] 0.3634926 0.5002413 0.6602156  0.03432499 0.6992927 0.7418657 0.5376907
#> [20,] 0.4606741 0.6383416 0.8263407  0.02503023 0.8544695 0.7582430 0.6481685
#> [21,] 0.7257688 0.7377738 0.6065859 -0.04741308 0.3735079 0.1670313 0.2486692
#> [22,] 0.7304648 0.7083172 0.4282645  0.05701565 0.4642010 0.6083587 0.5208640
#> [23,] 0.8061045 0.7699771 0.6940132  0.01127548 0.5456088 0.4034918 0.2611687
#> [24,] 0.6100742 0.4371820 0.3752521  0.02358693 0.4024848 0.4750942 0.5648366
#> [25,] 0.9585373 0.8704979 0.8244719  0.01595521 0.7618887 0.7016825 0.6555115
#>              V33        V34        V35        V36        V37        V38
#>  [1,] 0.26410719 0.25912818 0.19170209 0.04434549 0.06874573 0.14345131
#>  [2,] 0.55776113 0.34775631 0.15120288 0.20522893 0.29212604 0.41209910
#>  [3,] 0.09249875 0.05856083 0.10019276 0.21835945 0.22822965 0.28144903
#>  [4,] 0.14124509 0.31970822 0.37975513 0.23790066 0.22322813 0.25283127
#>  [5,] 0.19976479 0.07361665 0.03662393 0.09360314 0.22533031 0.23645454
#>  [6,] 0.23680751 0.37183490 0.41619521 0.34163785 0.33728109 0.41520729
#>  [7,] 0.38360104 0.37930227 0.35976495 0.43590005 0.45911061 0.48762559
#>  [8,] 0.29291124 0.23221504 0.31936432 0.07483190 0.18671433 0.31049135
#>  [9,] 0.24670453 0.03389414 0.09189890 0.14234498 0.22492983 0.30595436
#> [10,] 0.23024817 0.16773589 0.18558002 0.18108215 0.23691478 0.30541496
#> [11,] 0.71503190 0.62653034 0.63346712 0.45739568 0.27777923 0.30253239
#> [12,] 0.63258864 0.70973007 0.58013133 0.39549699 0.27340066 0.21011591
#> [13,] 0.17777562 0.20826217 0.24109649 0.18194184 0.06007335 0.07088302
#> [14,] 0.82736067 0.91116828 0.85766832 0.65820805 0.39404982 0.23041520
#> [15,] 0.55816578 0.53168271 0.60697548 0.63806252 0.50797351 0.23955753
#> [16,] 0.29120459 0.28800689 0.20509176 0.10917463 0.18648484 0.16784215
#> [17,] 0.36491827 0.32581392 0.14993502 0.10651843 0.04729640 0.04734449
#> [18,] 0.57318935 0.85595982 0.78894617 0.72443994 0.72755541 0.46059137
#> [19,] 0.53630485 0.69607261 0.79627056 0.84213600 0.73614662 0.71804989
#> [20,] 0.70252732 0.65582515 0.57545375 0.55694283 0.52633279 0.39326538
#> [21,] 0.31125445 0.32857029 0.18334594 0.15097510 0.24463768 0.25817944
#> [22,] 0.26103920 0.26966439 0.57595997 0.59623998 0.55488958 0.48198531
#> [23,] 0.18803341 0.25378167 0.41647018 0.31385674 0.09613895 0.06849919
#> [24,] 0.69201556 0.61524613 0.57919841 0.55849151 0.53675016 0.41392968
#> [25,] 0.60622310 0.61460951 0.63425842 0.60434973 0.54147733 0.43019522
#>              V39           V4        V40        V41        V42         V43
#>  [1,] 0.15270441  0.041255206 0.15879295 0.04252278 0.07760392 0.199193796
#>  [2,] 0.45249084  0.048894807 0.36750887 0.42984638 0.53240105 0.492680633
#>  [3,] 0.35381540  0.027470362 0.27707567 0.20169372 0.20363774 0.147636501
#>  [4,] 0.20214759  0.043890048 0.19038355 0.31960300 0.28528139 0.002897277
#>  [5,] 0.20603293  0.027606795 0.24385074 0.21952359 0.18296348 0.214670066
#>  [6,] 0.44313795  0.099615324 0.34256535 0.32551888 0.39647878 0.336806576
#>  [7,] 0.42582901  0.063558991 0.31166311 0.19604864 0.22190032 0.214941424
#>  [8,] 0.25999771  0.104211946 0.10613870 0.06761297 0.27136087 0.241339836
#>  [9,] 0.24537675  0.039001383 0.28749239 0.33077354 0.25183155 0.187094372
#> [10,] 0.41119575  0.079935929 0.43888263 0.36447716 0.31913134 0.296158387
#> [11,] 0.27480844  0.058311137 0.28679523 0.57742391 0.68040418 0.532020514
#> [12,] 0.29495106  0.028639445 0.41296511 0.34793863 0.33243159 0.278112554
#> [13,] 0.17133292  0.016076328 0.23660204 0.29461842 0.29248756 0.215920891
#> [14,] 0.23448966  0.031413612 0.34759575 0.32755862 0.24541002 0.307032666
#> [15,] 0.17621325  0.053263768 0.42581339 0.38749980 0.46367424 0.385977564
#> [16,] 0.10204854  0.034932092 0.15890057 0.12278678 0.08218181 0.071653492
#> [17,] 0.15201053  0.043737986 0.17577060 0.16090775 0.07131338 0.054085707
#> [18,] 0.33767645  0.075378812 0.30737438 0.21741408 0.27801840 0.106147074
#> [19,] 0.78186302  0.025667675 0.80283143 0.72704199 0.66514164 0.575326107
#> [20,] 0.29821731  0.027399247 0.23548410 0.18947599 0.16688878 0.162825041
#> [21,] 0.29855945 -0.057341968 0.25128931 0.23075943 0.18427206 0.128613636
#> [22,] 0.36645088  0.039234684 0.32179453 0.23939352 0.22513097 0.236959730
#> [23,] 0.06210505 -0.008628262 0.08267651 0.16133163 0.11320385 0.157824483
#> [24,] 0.38598358  0.020423499 0.57304754 0.71356396 0.67898339 0.676072595
#> [25,] 0.25473333  0.026241069 0.19528362 0.19351781 0.11019817 0.107110801
#>              V44        V45        V46         V47          V48        V49
#>  [1,] 0.22194036 0.18139105 0.17972701  0.15070464 0.0926363269 0.05527023
#>  [2,] 0.44190636 0.45474897 0.30196075  0.20538469 0.1930106658 0.10890317
#>  [3,] 0.18132227 0.14339088 0.11348895  0.12333487 0.1128887416 0.07941186
#>  [4,] 0.12108971 0.17982504 0.16832490  0.09359400 0.0645304292 0.01603270
#>  [5,] 0.17084165 0.09502386 0.06701202  0.13319407 0.0969551081 0.05277711
#>  [6,] 0.26734662 0.19238213 0.08539544  0.05447794 0.0141324236 0.02642370
#>  [7,] 0.16613987 0.22065580 0.16926211  0.12820398 0.1136710334 0.07670366
#>  [8,] 0.04678066 0.14478271 0.12048918  0.05245695 0.0564959302 0.04064664
#>  [9,] 0.22277323 0.13159245 0.07864692  0.10682002 0.0918083476 0.06103745
#> [10,] 0.21721182 0.20645919 0.15343357  0.11428804 0.0661406718 0.05256730
#> [11,] 0.50589822 0.60350323 0.54758204  0.32171144 0.1796239246 0.07971906
#> [12,] 0.18222295 0.15978839 0.11490396  0.12179478 0.1322019552 0.06469375
#> [13,] 0.12704244 0.10172104 0.10940482  0.05451590 0.0334709225 0.02838264
#> [14,] 0.23358605 0.13753740 0.08276146  0.05362574 0.0397311200 0.01671267
#> [15,] 0.25724553 0.14439946 0.09192635  0.04283193 0.0504744896 0.02410105
#> [16,] 0.12034136 0.12119025 0.08292097  0.09663282 0.0701293192 0.03439334
#> [17,] 0.10716658 0.07472210 0.06130051  0.07308646 0.0377862222 0.02772712
#> [18,] 0.25091984 0.17247809 0.15391084  0.11346176 0.0735355444 0.03811466
#> [19,] 0.44257848 0.32371956 0.26711161  0.21684280 0.1426013072 0.08206680
#> [20,] 0.19630426 0.17926549 0.12640924  0.08679878 0.0476183415 0.02169387
#> [21,] 0.17939980 0.15472738 0.12616921  0.01914790 0.0091801207 0.01814177
#> [22,] 0.20867315 0.17365980 0.07319144  0.08672359 0.0650454190 0.04028953
#> [23,] 0.11677132 0.02981359 0.01639582 -0.01502485 0.0009047531 0.01150272
#> [24,] 0.43918974 0.32633321 0.24862721  0.14171255 0.0921706742 0.02006644
#> [25,] 0.07307478 0.05219166 0.04720936  0.02171014 0.0168506727 0.00857566
#>                V5         V50          V51           V52         V53
#>  [1,]  0.04755761 0.010168493  0.017092799  1.029843e-02 0.001437636
#>  [2,]  0.12254650 0.033686681  0.028873775  1.871021e-02 0.009921904
#>  [3,]  0.11650992 0.025139276  0.023794692  1.805908e-02 0.013946407
#>  [4,]  0.12522472 0.016250460  0.017014246  1.928765e-02 0.021612619
#>  [5,]  0.09526346 0.030932262  0.019399644  1.766260e-02 0.020211654
#>  [6,]  0.14723692 0.007601422  0.006455046  2.731801e-03 0.009217428
#>  [7,]  0.06871203 0.037487809  0.019042975  1.134204e-02 0.007380514
#>  [8,]  0.16012811 0.037367371  0.026098470  2.328591e-02 0.022620518
#>  [9,]  0.06202679 0.019015960  0.013536483  1.517656e-02 0.013568593
#> [10,]  0.12019237 0.025697311  0.014996571  8.844121e-03 0.014752234
#> [11,]  0.08980592 0.033887194  0.040691419  3.801368e-02 0.023508818
#> [12,]  0.03499936 0.016110446  0.011000527  8.862208e-03 0.007041258
#> [13,]  0.03245104 0.009446072  0.005078665  7.794057e-03 0.003550378
#> [14,]  0.05951855 0.015904549  0.013250119  1.412232e-02 0.017697059
#> [15,]  0.06979464 0.007809122  0.003590150  6.559567e-03 0.011973517
#> [16,]  0.06618784 0.019128849  0.011065892  9.050391e-03 0.008772068
#> [17,]  0.07200388 0.014023857  0.016269768  4.817082e-03 0.008999969
#> [18,]  0.15576140 0.011814785  0.014852151  2.601229e-03 0.012923555
#> [19,]  0.08140251 0.026510815  0.017408127  7.865745e-03 0.015065673
#> [20,]  0.03295820 0.010644714  0.008209699  7.240383e-03 0.012880158
#> [21,] -0.03662058 0.004159856 -0.006185799 -2.175185e-03 0.003638934
#> [22,]  0.05619159 0.020499507  0.008960297  1.393165e-02 0.005799384
#> [23,]  0.02801651 0.012131613  0.006404478  1.313112e-02 0.004278168
#> [24,]  0.01050972 0.024010955  0.007103003 -8.992895e-05 0.004745053
#> [25,]  0.04699453 0.011287150  0.003424237  5.544501e-03 0.006904612
#>               V54           V55           V56           V57          V58
#>  [1,] 0.004885830  0.0033863037  0.0012171086  0.0046406261  0.005203328
#>  [2,] 0.011889776  0.0149100833  0.0106339312  0.0091244907  0.009125612
#>  [3,] 0.023165777  0.0177843605  0.0058147330  0.0070885869  0.009791459
#>  [4,] 0.029150120  0.0145719318  0.0128565235  0.0119638879  0.009733235
#>  [5,] 0.009520753  0.0035929077  0.0102214523  0.0129789410  0.012558436
#>  [6,] 0.024084352  0.0122292431  0.0148386668  0.0153390184  0.013377052
#>  [7,] 0.015738319  0.0071959966  0.0140869510  0.0074658273  0.012170052
#>  [8,] 0.020871513  0.0178506011  0.0082375806  0.0030762112  0.015761700
#>  [9,] 0.011380385  0.0067537941  0.0153887457  0.0084238294  0.003447869
#> [10,] 0.007327812  0.0070512793  0.0094723852  0.0065068414  0.006939078
#> [11,] 0.015522218  0.0109260104  0.0099147130  0.0084947260  0.011230795
#> [12,] 0.010816888  0.0080917720  0.0065075168  0.0037437941  0.008817304
#> [13,] 0.003752579  0.0056286294  0.0032907405  0.0025868368  0.003452800
#> [14,] 0.013606709  0.0059750645  0.0112906651  0.0186809565  0.007546652
#> [15,] 0.015158302  0.0119787742  0.0140200714  0.0113546602  0.006721196
#> [16,] 0.007163616  0.0062044054  0.0099085623  0.0069687491  0.003543780
#> [17,] 0.009444965  0.0080600235  0.0074464208  0.0027717597  0.003097643
#> [18,] 0.007029876  0.0171575884  0.0078674641  0.0149978144  0.004942079
#> [19,] 0.010794265  0.0113161720  0.0118532068  0.0045520538  0.015410052
#> [20,] 0.011524709  0.0098428999  0.0062365408  0.0067548579  0.006597164
#> [21,] 0.006771539 -0.0029989349 -0.0007878506 -0.0004297222  0.002491707
#> [22,] 0.014039793  0.0115221918  0.0071163505  0.0114345218  0.007249460
#> [23,] 0.005217242  0.0047523260  0.0056616368  0.0028221124 -0.000925852
#> [24,] 0.002309717 -0.0003931091  0.0079835686  0.0097547703  0.004703534
#> [25,] 0.007445925  0.0068850175  0.0053039545  0.0050811253  0.004317493
#>                 V59          V6         V60         V7         V8           V9
#>  [1,]  0.0037939737  0.08635806 0.003526101 0.11990349 0.14434044  0.221524481
#>  [2,]  0.0078043590  0.09722232 0.003969303 0.09894193 0.12520729  0.250535126
#>  [3,]  0.0092139292  0.18055991 0.005544851 0.18081559 0.15656881  0.189342265
#>  [4,]  0.0006075675  0.17491834 0.002931638 0.16457594 0.22239210  0.343971096
#>  [5,]  0.0106609699  0.14159230 0.009896033 0.13229360 0.15141911  0.207724990
#>  [6,]  0.0156369402  0.15301710 0.011615662 0.12144250 0.10795187  0.077927155
#>  [7,]  0.0083496746  0.07209029 0.008982787 0.09076766 0.21090363  0.347843585
#>  [8,]  0.0146864278  0.15232179 0.010678673 0.14890592 0.09095713  0.111987054
#>  [9,]  0.0056000237  0.09086747 0.002840762 0.12988743 0.10922956  0.155727141
#> [10,]  0.0093648210  0.12315848 0.006973764 0.14957519 0.10951983  0.159045358
#> [11,]  0.0083449844  0.16166361 0.005549766 0.12837277 0.16352321  0.239645512
#> [12,]  0.0061310446  0.06102782 0.003627640 0.09927089 0.17258473  0.306869655
#> [13,]  0.0015403778  0.04656951 0.003017573 0.06407012 0.06289269  0.106906458
#> [14,]  0.0073663888  0.06050737 0.011186862 0.09422393 0.07796633  0.035317502
#> [15,]  0.0086469536  0.11935059 0.012572441 0.10885453 0.12797058  0.050793151
#> [16,]  0.0052131818  0.09957467 0.006021105 0.13392966 0.10434039  0.124679688
#> [17,]  0.0026937116  0.08001394 0.002383988 0.09302816 0.05221923  0.049623254
#> [18,]  0.0159918315  0.21073473 0.005561171 0.20699115 0.17498937  0.160489333
#> [19,]  0.0200804407  0.09643832 0.011984831 0.10937610 0.12037235 -0.002301581
#> [20,]  0.0088234592  0.05189209 0.004363491 0.06316369 0.04765504  0.056051682
#> [21,] -0.0003985710 -0.03653422 0.001772322 0.03819729 0.10618696  0.069483520
#> [22,]  0.0102656321  0.16911051 0.006852626 0.18378139 0.24688971  0.351140593
#> [23,]  0.0003122247  0.12097693 0.001338898 0.13047964 0.09823118  0.145538162
#> [24,]  0.0085257950  0.06412986 0.006348233 0.07576820 0.09725742  0.080611492
#> [25,]  0.0037655218  0.08617452 0.004156238 0.10508460 0.09651961  0.137923828
#> 
#> $cl
#>  [1] M M M M M M M M M M M M R R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478