Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                  V1        V10         V11        V12        V13        V14
#>  [1,]  0.0916657734 0.13292298  0.22058703 0.23795015 0.28128129 0.30473632
#>  [2,]  0.0098706597 0.20778204  0.07181316 0.01129452 0.21977913 0.34189421
#>  [3,]  0.0407414535 0.25228029  0.35225025 0.41359324 0.46937450 0.60536790
#>  [4,]  0.0457761157 0.19281954  0.19402290 0.27391661 0.33872850 0.31254377
#>  [5,]  0.0869969976 0.38933507  0.47570890 0.52128360 0.63545000 0.71062006
#>  [6,]  0.0261378965 0.28440380  0.31227579 0.30097403 0.21528350 0.11644549
#>  [7,]  0.0145467873 0.17198158  0.22306814 0.27661716 0.33392336 0.40060594
#>  [8,]  0.0286980086 0.33814230  0.35454075 0.38272263 0.43545291 0.38373157
#>  [9,]  0.0348995922 0.30622067  0.35703903 0.37023087 0.38185667 0.36249059
#> [10,]  0.0549484605 0.45631427  0.40041911 0.23028448 0.22315517 0.30739031
#> [11,]  0.0432868001 0.27845804  0.29765849 0.26185393 0.21040483 0.10867519
#> [12,]  0.0320303734 0.13494072  0.21064145 0.28258387 0.30055800 0.35080528
#> [13,]  0.0252373181 0.33045479  0.26399831 0.29122570 0.24650791 0.20921050
#> [14,]  0.0387291565 0.19698084  0.32003749 0.35696642 0.34044781 0.32029316
#> [15,]  0.0339985989 0.04204782  0.07982786 0.24568683 0.13645846 0.11696295
#> [16,]  0.0622605834 0.12240207  0.25949083 0.44041168 0.39577968 0.17860394
#> [17,]  0.0233708959 0.15501772  0.10890186 0.12735679 0.10305382 0.10241964
#> [18,]  0.0190906866 0.10929503  0.14064637 0.16661082 0.19233862 0.22825441
#> [19,]  0.0284973695 0.09659725  0.06062789 0.19052839 0.20062882 0.13698429
#> [20,]  0.0177936651 0.07678077  0.09121926 0.09025052 0.18811776 0.31639057
#> [21,]  0.0210714337 0.16749380  0.24020108 0.17226476 0.15290893 0.16465400
#> [22,]  0.0264426300 0.24942243  0.22761261 0.20443946 0.28436957 0.37647730
#> [23,]  0.0293000000 0.08600000  0.04140000 0.04720000 0.08350000 0.09380000
#> [24,]  0.0166434215 0.12727397  0.11162864 0.09734671 0.09637605 0.12367165
#> [25,] -0.0004801908 0.03540215 -0.01503130 0.16025606 0.13745121 0.08112928
#>             V15        V16        V17       V18        V19           V2
#>  [1,] 0.3834319 0.32747773 0.40219731 0.2135173 0.30028818 0.0817235703
#>  [2,] 0.4801822 0.62965329 0.67686763 0.7217666 0.81414837 0.0019064770
#>  [3,] 0.8081831 0.82719478 0.78490343 0.8989301 1.01132943 0.0609080457
#>  [4,] 0.3124867 0.41090877 0.40466725 0.1611814 0.34654915 0.0996816690
#>  [5,] 0.8314981 0.89945845 0.88703274 0.8913325 0.93409556 0.0841141689
#>  [6,] 0.1273408 0.13646089 0.17823327 0.2986104 0.44367782 0.0386858244
#>  [7,] 0.4470470 0.57347425 0.66293578 0.7856006 0.89316191 0.0308720160
#>  [8,] 0.2879462 0.31363227 0.30483909 0.2691227 0.43162572 0.0389202227
#>  [9,] 0.3364841 0.38351012 0.46705385 0.5426489 0.66358334 0.0360567539
#> [10,] 0.3293732 0.28818346 0.22302377 0.3061998 0.51275486 0.0748943613
#> [11,] 0.1003906 0.16913121 0.19445491 0.2483555 0.20566267 0.0415748654
#> [12,] 0.4419426 0.54091872 0.67169974 0.7674608 0.85498763 0.0392847658
#> [13,] 0.1145210 0.14892435 0.16605963 0.2344110 0.17877338 0.0360679508
#> [14,] 0.2303729 0.12405322 0.25781079 0.2670997 0.31535916 0.0451599130
#> [15,] 0.1125284 0.16344522 0.14517064 0.1128126 0.17501773 0.0359883617
#> [16,] 0.2563409 0.39388266 0.62854411 0.7381823 0.63016686 0.0521849997
#> [17,] 0.1046451 0.15079372 0.22027465 0.4684969 0.53828229 0.0270734467
#> [18,] 0.2405956 0.30043913 0.36877618 0.3526236 0.30276458 0.0244491245
#> [19,] 0.1717314 0.21628063 0.20286669 0.1162790 0.11897156 0.0456213819
#> [20,] 0.5000404 0.65591235 0.70831439 0.7303809 0.79501289 0.0160338730
#> [21,] 0.1666046 0.17714087 0.25343860 0.1198251 0.38507733 0.0035385164
#> [22,] 0.4305333 0.51276622 0.57326545 0.5837280 0.56176178 0.0570489191
#> [23,] 0.1466000 0.08090000 0.11790000 0.2179000 0.33260000 0.0644000000
#> [24,] 0.1139458 0.15836003 0.20403192 0.2513095 0.24904118 0.0145676114
#> [25,] 0.1067175 0.08581406 0.07526994 0.2090805 0.04977791 0.0004201504
#>              V20        V21        V22        V23       V24        V25
#>  [1,] 0.33301603 0.31112605 0.31333357 0.39546783 0.2812739 0.41352004
#>  [2,] 0.86935210 0.96900961 0.98251746 0.89415795 0.6513769 0.47752746
#>  [3,] 0.92179628 0.90573488 0.94752246 0.93614912 0.8109452 0.66168314
#>  [4,] 0.39006212 0.52912991 0.53834606 0.52245375 0.6340325 0.52585761
#>  [5,] 0.98019078 0.95477248 0.85851504 0.79144955 0.6479550 0.49651260
#>  [6,] 0.56112678 0.69836515 0.79241495 0.78606752 0.8741280 0.94311343
#>  [7,] 0.97114006 1.00262581 0.99770427 0.94832069 0.8610287 0.76998319
#>  [8,] 0.62854541 0.64598210 0.62145954 0.59315714 0.5932964 0.65426850
#>  [9,] 0.77844502 0.86243852 0.89137940 0.88586946 0.9091225 0.93098931
#> [10,] 0.86958243 0.85621413 0.39911606 0.14601926 0.1712315 0.29854651
#> [11,] 0.15189165 0.32906196 0.61497240 0.78301986 0.8484103 0.92089874
#> [12,] 0.90223729 0.89670514 0.86950015 0.82124433 0.7585271 0.63060142
#> [13,] 0.13590311 0.03604151 0.14978968 0.39227036 0.5463249 0.66567461
#> [14,] 0.40170250 0.56024829 0.57553212 0.54076353 0.5649839 0.57836207
#> [15,] 0.11573884 0.19181388 0.02904213 0.23336462 0.2335772 0.02929629
#> [16,] 0.42874292 0.23536502 0.17181713 0.22213204 0.1777663 0.25275840
#> [17,] 0.58053956 0.60121068 0.67460349 0.68799175 0.6283783 0.40895319
#> [18,] 0.33976453 0.35694176 0.38653104 0.48718298 0.6529470 0.71108315
#> [19,] 0.14454010 0.21488640 0.15130410 0.22698618 0.2437368 0.17769529
#> [20,] 0.84911378 0.97378902 0.97455976 0.94152869 0.9491152 0.89656632
#> [21,] 0.56735813 0.33216755 0.34350034 0.99924348 0.9138348 0.47192232
#> [22,] 0.57201322 0.55969588 0.54284456 0.57549888 0.5931537 0.58730331
#> [23,] 0.32580000 0.21110000 0.23020000 0.33610000 0.4259000 0.46090000
#> [24,] 0.45808301 0.58349658 0.79186263 0.96828108 0.9116315 0.87857157
#> [25,] 0.06383254 0.22482947 0.13292715 0.05809966 0.1364287 0.02861110
#>             V26       V27       V28       V29          V3       V30       V31
#>  [1,] 0.4567740 0.4727706 0.7346784 0.7627490 0.065791135 0.6589858 0.7472336
#>  [2,] 0.5090246 0.6055171 0.7751034 0.7821395 0.024276297 0.5833653 0.2951256
#>  [3,] 0.6040594 0.4638737 0.3177952 0.4675974 0.066178364 0.3403654 0.2696652
#>  [4,] 0.4772479 0.6978434 0.8742855 0.8799464 0.066586377 0.8989831 0.8462283
#>  [5,] 0.5017300 0.5219932 0.5116248 0.3506873 0.060552977 0.2933530 0.1839875
#>  [6,] 1.0322326 1.0109429 0.8686095 0.7729695 0.032359979 0.7292618 0.6192007
#>  [7,] 0.7672812 0.7706565 0.6948244 0.5527375 0.037624819 0.4550591 0.3354295
#>  [8,] 0.7811712 0.8416499 0.8593877 0.7942375 0.056516228 0.7171923 0.6744963
#>  [9,] 0.9830530 0.9855921 0.9020459 0.7049550 0.027029865 0.5375066 0.4276657
#> [10,] 0.6073484 0.6125909 0.7048859 0.7027088 0.060657894 0.8012452 0.4925766
#> [11,] 0.9655292 0.9684251 1.0039505 0.8472805 0.022095864 0.6644227 0.7303146
#> [12,] 0.4996498 0.3936999 0.2939629 0.3133664 0.046474218 0.3369239 0.2863751
#> [13,] 0.7981797 0.9311414 0.9718611 0.9206651 0.027985575 0.9420048 0.9244784
#> [14,] 0.7555159 0.9627939 1.0105042 0.8693273 0.052616751 0.7310673 0.6715429
#> [15,] 0.1933645 0.4765542 0.6931256 0.6781699 0.033410632 0.6369984 0.7149269
#> [16,] 0.2572167 0.3070369 0.4595366 0.7123923 0.088646807 0.7428123 0.8491333
#> [17,] 0.3367423 0.4357800 0.6940135 0.8534621 0.014550038 0.8095338 0.7861761
#> [18,] 0.6908435 0.6955287 0.8212789 0.9059815 0.028045842 0.8489063 0.7906989
#> [19,] 0.3412045 0.6261070 0.7962346 0.7914631 0.016904884 0.7561285 0.7833211
#> [20,] 0.8841298 0.8175029 0.6113206 0.4386988 0.020179626 0.4523264 0.4444228
#> [21,] 0.5406252 0.6437887 0.7543207 0.6060047 0.053892076 0.4251277 0.2948415
#> [22,] 0.6385345 0.7039414 0.6073508 0.4205686 0.068618016 0.4012544 0.4631573
#> [23,] 0.2606000 0.0874000 0.2862000 0.5606000 0.039000000 0.8344000 0.8096000
#> [24,] 0.8924177 0.8390173 0.7984773 0.7912250 0.024358481 0.7191661 0.6819494
#> [25,] 0.1476469 0.3534354 0.4878670 0.5256458 0.004560854 0.5334616 0.6522065
#>              V32        V33       V34       V35        V36        V37
#>  [1,] 0.68923602 0.54974736 0.5858854 0.6079783 0.37292893 0.10473018
#>  [2,] 0.30831619 0.32744942 0.2052624 0.1501697 0.10147346 0.08372295
#>  [3,] 0.09188609 0.03195808 0.1235116 0.1887844 0.16804047 0.15035804
#>  [4,] 0.72345993 0.38994066 0.4098525 0.5039634 0.31291833 0.24397466
#>  [5,] 0.16732880 0.18601168 0.1651818 0.1857860 0.01573235 0.09127979
#>  [6,] 0.40724400 0.24376895 0.2214522 0.1725908 0.08878423 0.16132518
#>  [7,] 0.21522135 0.08963378 0.1472174 0.2577858 0.29487534 0.24526582
#>  [8,] 0.67287603 0.53747883 0.3403121 0.1236147 0.21262558 0.24162892
#>  [9,] 0.34020807 0.27951777 0.2012434 0.1218767 0.20573904 0.24369428
#> [10,] 0.47240796 0.66721774 0.7041984 0.6837103 0.59650781 0.59573210
#> [11,] 0.69547932 0.50297562 0.1784575 0.1015866 0.11474722 0.11415688
#> [12,] 0.30520314 0.36256832 0.3791203 0.3543972 0.36313381 0.38716012
#> [13,] 0.85453035 0.73270669 0.6423198 0.6485419 0.69955356 0.72158239
#> [14,] 0.48690819 0.26739294 0.2187243 0.1085186 0.04523031 0.24450072
#> [15,] 0.54187298 0.64938272 0.8663187 0.9527380 0.96260276 0.77104620
#> [16,] 0.85429775 0.68328487 0.6851288 0.7288781 0.81987651 0.90699227
#> [17,] 0.66534896 0.57326867 0.5214159 0.4976180 0.51803806 0.37926404
#> [18,] 0.72064680 0.70474978 0.6690217 0.6057140 0.55782763 0.51282917
#> [19,] 0.64675737 0.68588772 0.8453403 0.9495911 0.93076640 0.72570767
#> [20,] 0.34553262 0.34236874 0.2902225 0.1514463 0.06947790 0.14099949
#> [21,] 0.19886445 0.12974433 0.3028784 0.2627583 0.35854213 0.30699297
#> [22,] 0.44907586 0.36205339 0.2378665 0.3213827 0.45480983 0.53841038
#> [23,] 0.72500000 0.80480000 0.9435000 1.0000000 0.89600000 0.55160000
#> [24,] 0.56656797 0.54976629 0.5732531 0.5596188 0.44433201 0.36566475
#> [25,] 0.63208340 0.68366160 0.9021857 1.0137082 0.95776110 0.84745270
#>              V38        V39         V4        V40         V41        V42
#>  [1,] 0.29680287 0.26770384 0.07107686 0.11437236 0.511946465 0.77349259
#>  [2,] 0.07852956 0.06660587 0.06035192 0.11192480 0.224189398 0.24057451
#>  [3,] 0.08364751 0.30317611 0.05187422 0.29879380 0.150244835 0.13581153
#>  [4,] 0.31558598 0.24026133 0.04434780 0.16858779 0.593679065 0.64552648
#>  [5,] 0.16148070 0.15915025 0.07411374 0.18483257 0.120836395 0.17579197
#>  [6,] 0.16347073 0.13152195 0.03679921 0.14442786 0.068266453 0.16170883
#>  [7,] 0.24837822 0.27983422 0.04378271 0.22557651 0.253941039 0.29763896
#>  [8,] 0.37437369 0.55720095 0.04947904 0.49197279 0.392871461 0.47792431
#>  [9,] 0.28411299 0.30191307 0.04198610 0.29016546 0.156953270 0.23565535
#> [10,] 0.65184809 0.56657792 0.09861239 0.41347609 0.399599715 0.37231756
#> [11,] 0.15251061 0.18959020 0.03539414 0.05744136 0.146207474 0.04660938
#> [12,] 0.38205984 0.33875283 0.06078630 0.32711579 0.309111729 0.29049706
#> [13,] 0.58892971 0.48346730 0.04480667 0.52878158 0.488665724 0.42991369
#> [14,] 0.20869820 0.20267214 0.08618326 0.27773154 0.296830099 0.35247768
#> [15,] 0.66461569 0.63112573 0.02757027 0.66622348 0.595029526 0.51201188
#> [16,] 1.01913088 0.88894994 0.02509468 0.72632550 0.724590582 0.62317412
#> [17,] 0.41504811 0.41583159 0.03718061 0.35109978 0.241022917 0.13781875
#> [18,] 0.38719383 0.29604331 0.03486297 0.22785091 0.187987348 0.19790187
#> [19,] 0.59779829 0.54576806 0.01404433 0.61257029 0.531010823 0.48524744
#> [20,] 0.12492702 0.13596411 0.04000914 0.16371911 0.118212946 0.04816518
#> [21,] 0.21263082 0.25944678 0.04372982 0.12749958 0.005984291 0.00897615
#> [22,] 0.41903890 0.39375583 0.05920598 0.42023063 0.355832349 0.30823158
#> [23,] 0.30370000 0.23380000 0.01730000 0.23820000 0.331800000 0.38210000
#> [24,] 0.20529435 0.08386530 0.03307761 0.15786495 0.159839390 0.13731390
#> [25,] 0.68058580 0.65793069 0.04564676 0.64689614 0.527670374 0.43839092
#>              V43        V44         V45         V46          V47         V48
#>  [1,] 0.69208161 0.52636930  0.58353479  0.70251963  0.520103413 0.301868554
#>  [2,] 0.11259009 0.19101620  0.21752814  0.17383554  0.098186601 0.038099731
#>  [3,] 0.30808158 0.21579666  0.09285359  0.12908379  0.094798746 0.037979325
#>  [4,] 0.49532367 0.42510584  0.53105686  0.57377880  0.394054105 0.224429906
#>  [5,] 0.16690079 0.14622664  0.15690110  0.09970174  0.062057810 0.078014488
#>  [6,] 0.24518277 0.23272810  0.20484773  0.19479929  0.149656388 0.068497894
#>  [7,] 0.22550948 0.18837375  0.11848943  0.10572670  0.101058421 0.103948108
#>  [8,] 0.45692970 0.46108968  0.54753662  0.36047797  0.242176521 0.217350222
#>  [9,] 0.31217083 0.29931475  0.25383142  0.13205096  0.097020487 0.081486410
#> [10,] 0.32502739 0.24830881  0.28818564  0.23295677  0.103289156 0.100996949
#> [11,] 0.07838548 0.13303053  0.08705160  0.07886703  0.108644612 0.107802562
#> [12,] 0.24077308 0.18743377  0.11569197  0.08024849  0.072710701 0.042683703
#> [13,] 0.45322649 0.52650155  0.51074185  0.38127634  0.285029343 0.222746985
#> [14,] 0.26522544 0.15222430  0.18805850  0.21521847  0.228359315 0.167299997
#> [15,] 0.37736284 0.22306506  0.14357389  0.07558938  0.080661749 0.075932982
#> [16,] 0.47841735 0.36536991  0.10053035  0.08361761  0.131943998 0.118510096
#> [17,] 0.20242234 0.17319754  0.09515598  0.06274527  0.053492722 0.027963267
#> [18,] 0.20189344 0.20069879  0.15487841  0.12187767  0.074975805 0.055916854
#> [19,] 0.35487653 0.21898081  0.18151253  0.10027023  0.110069786 0.096996770
#> [20,] 0.05218003 0.11085584  0.09268744  0.03686088  0.055064169 0.044143724
#> [21,] 0.08168381 0.10829722  0.04344118 -0.01913666 -0.007751512 0.001909297
#> [22,] 0.19405039 0.15301317  0.13042539  0.09205261  0.087593966 0.070603201
#> [23,] 0.15750000 0.22280000  0.15820000  0.14330000  0.163400000 0.113300000
#> [24,] 0.15154488 0.03888586 -0.02329825  0.01811722  0.030263388 0.043328361
#> [25,] 0.29853142 0.16444082  0.11084327  0.07234111  0.075082847 0.073145152
#>               V49         V5         V50         V51         V52         V53
#>  [1,] 0.146612464 0.06420003 0.045861582 0.091157823 0.068448463 0.031527468
#>  [2,] 0.032442705 0.08569016 0.006665597 0.005248358 0.006564127 0.002735210
#>  [3,] 0.033455232 0.01637977 0.018217179 0.038402892 0.011673259 0.006160136
#>  [4,] 0.133574187 0.06032740 0.061717923 0.066455494 0.038615019 0.016280607
#>  [5,] 0.069287522 0.13288986 0.027311674 0.029620425 0.023684169 0.019730917
#>  [6,] 0.044676919 0.04178433 0.015996631 0.019604427 0.013291899 0.007149693
#>  [7,] 0.059786178 0.12713559 0.021220499 0.022585099 0.025667962 0.019219492
#>  [8,] 0.118570949 0.09076287 0.034995824 0.025365654 0.017322699 0.010926973
#>  [9,] 0.051817222 0.10160571 0.017014902 0.019564873 0.008543332 0.014028469
#> [10,] 0.052172801 0.12616948 0.020702037 0.018482581 0.021207034 0.017184909
#> [11,] 0.062036455 0.04692585 0.025129353 0.016276622 0.014286683 0.004285566
#> [12,] 0.027770264 0.09280631 0.014394367 0.008944765 0.012337286 0.010854964
#> [13,] 0.121598461 0.06190928 0.036111420 0.020594537 0.018032526 0.007674551
#> [14,] 0.072385228 0.10394581 0.018936745 0.022738692 0.014235538 0.009030969
#> [15,] 0.055563165 0.07319442 0.017659551 0.005573926 0.008012316 0.016230074
#> [16,] 0.056399891 0.03763451 0.021433744 0.012760801 0.012245759 0.018059295
#> [17,] 0.030171459 0.04816153 0.028163754 0.007188373 0.013850825 0.013322792
#> [18,] 0.030128399 0.04284297 0.015967708 0.009964980 0.007945516 0.007128900
#> [19,] 0.050310842 0.03451803 0.018727036 0.012595838 0.008264455 0.019335705
#> [20,] 0.027855743 0.06728479 0.011887457 0.016360306 0.007267423 0.010078260
#> [21,] 0.004284393 0.05676521 0.004350354 0.031076784 0.044459166 0.016484437
#> [22,] 0.044619666 0.05393377 0.016180959 0.015072629 0.012286679 0.008750491
#> [23,] 0.056700000 0.04760000 0.013300000 0.017000000 0.003500000 0.005200000
#> [24,] 0.011398863 0.07060673 0.006547250 0.002085075 0.003080264 0.009928643
#> [25,] 0.028017628 0.02352481 0.022197197 0.024323012 0.002742713 0.003927066
#>                 V54           V55         V56          V57         V58
#>  [1,]  0.0291468468  0.0226368012 0.007592903 0.0186740706 0.019939779
#>  [2,]  0.0006840464 -0.0005845525 0.002260741 0.0016415800 0.006159064
#>  [3,]  0.0257136610  0.0023706951 0.009083235 0.0130406171 0.012483493
#>  [4,]  0.0257468737  0.0164821490 0.015747252 0.0091422830 0.017629290
#>  [5,]  0.0183096031  0.0217250464 0.001956142 0.0004727395 0.017942343
#>  [6,]  0.0052404862  0.0021031294 0.002714624 0.0036563352 0.003418621
#>  [7,]  0.0273415501  0.0107612950 0.005966970 0.0071358844 0.012967443
#>  [8,]  0.0111118267  0.0145918797 0.008982308 0.0107066742 0.009581847
#>  [9,]  0.0095405863  0.0065634570 0.006906537 0.0048720961 0.008373670
#> [10,]  0.0159859415  0.0101516315 0.010036798 0.0068391184 0.009704340
#> [11,]  0.0075034133  0.0054208016 0.009315666 0.0116039511 0.012300228
#> [12,]  0.0135023128  0.0118689763 0.012884903 0.0090221419 0.009336350
#> [13,]  0.0067063658  0.0055377222 0.007195833 0.0066450713 0.003880108
#> [14,]  0.0072110333  0.0053716886 0.005872662 0.0053552271 0.008147419
#> [15,]  0.0158518685  0.0079310288 0.006210078 0.0129914195 0.008437142
#> [16,]  0.0019214374  0.0015372603 0.000469076 0.0008253531 0.016854412
#> [17,]  0.0087311307  0.0068288061 0.003998456 0.0057230773 0.006305178
#> [18,]  0.0082376508  0.0095142631 0.005964143 0.0080667783 0.005869916
#> [19,]  0.0138289065  0.0092186814 0.009457891 0.0068773713 0.007765624
#> [20,]  0.0083005257  0.0055315440 0.006655048 0.0029481514 0.003097125
#> [21,] -0.0019878180  0.0089061786 0.007807485 0.0081184852 0.018949540
#> [22,]  0.0146847764  0.0127145169 0.008493811 0.0111338792 0.009649121
#> [23,]  0.0083000000  0.0078000000 0.007500000 0.0105000000 0.016000000
#> [24,]  0.0103851263  0.0102514706 0.006731381 0.0083477717 0.005522752
#> [25,]  0.0057856636  0.0116833076 0.009503544 0.0039221548 0.001774790
#>               V59         V6         V60         V7         V8           V9
#>  [1,] 0.020593053 0.12024951 0.008717287 0.08130773 0.15999545  0.131177300
#>  [2,] 0.001927057 0.12720518 0.002915713 0.15401022 0.19723902  0.248267839
#>  [3,] 0.010317308 0.14747575 0.014032095 0.21249929 0.28252258  0.311714404
#>  [4,] 0.013049823 0.13035756 0.009811031 0.15288197 0.19561241  0.262450901
#>  [5,] 0.014174700 0.17444321 0.014625197 0.19834456 0.12961133  0.218517320
#>  [6,] 0.004176211 0.05756622 0.007625335 0.07974442 0.13410744  0.208524620
#>  [7,] 0.008881492 0.19236417 0.003886745 0.15946905 0.11105499  0.136916686
#>  [8,] 0.007376556 0.10641262 0.004320844 0.11666427 0.15146388  0.249355570
#>  [9,] 0.007063972 0.16281962 0.004727156 0.15054358 0.16069629  0.246214557
#> [10,] 0.009997662 0.09367710 0.008563975 0.12794325 0.24696517  0.470734271
#> [11,] 0.009585068 0.12440684 0.006293326 0.14056766 0.14351638  0.170720880
#> [12,] 0.010774250 0.09739836 0.007814265 0.10096295 0.08958685  0.114386138
#> [13,] 0.002838912 0.08657795 0.001828382 0.11269727 0.13654081  0.293416199
#> [14,] 0.004234097 0.11124628 0.005679961 0.10934411 0.11325933  0.150459547
#> [15,] 0.011198844 0.11746628 0.006564007 0.10011077 0.05728017  0.006000511
#> [16,] 0.011477112 0.07639450 0.007314526 0.06413478 0.08264985  0.188809823
#> [17,] 0.005414062 0.10315999 0.006447596 0.12189699 0.12864889  0.124228009
#> [18,] 0.010203196 0.08591751 0.006037251 0.10738075 0.10100916  0.099815900
#> [19,] 0.008508144 0.05489913 0.008487226 0.05597403 0.08284257  0.080912369
#> [20,] 0.003907405 0.09467296 0.003083738 0.12490267 0.08037231  0.091874271
#> [21,] 0.011233711 0.15909622 0.005938796 0.13871393 0.22116816  0.153065101
#> [22,] 0.006795291 0.08350624 0.004742544 0.10132537 0.14338793  0.191743117
#> [23,] 0.009500000 0.08160000 0.001100000 0.09930000 0.03150000  0.073600000
#> [24,] 0.004397485 0.11254272 0.007305284 0.11513462 0.11401470  0.116884203
#> [25,] 0.006420317 0.07550217 0.005542889 0.04675600 0.03625821 -0.054009791
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406