Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.108355274 0.18282293 0.27716426 0.23324561 0.28541430 0.3385979
#>  [2,] 0.026884430 0.28686943 0.35580269 0.40213858 0.40258912 0.4603673
#>  [3,] 0.031518656 0.40430031 0.36083803 0.23964296 0.25623315 0.3025357
#>  [4,] 0.026702317 0.21828549 0.26738494 0.28313210 0.23124150 0.1602556
#>  [5,] 0.052808884 0.18670858 0.20145767 0.19603895 0.29840347 0.3580320
#>  [6,] 0.042572991 0.29135431 0.31764004 0.30679637 0.28366033 0.2926947
#>  [7,] 0.037909697 0.27303706 0.34943014 0.44069266 0.38456630 0.3965107
#>  [8,] 0.010803021 0.13168366 0.20111061 0.22697594 0.21233537 0.1379743
#>  [9,] 0.045449978 0.16216751 0.23120645 0.24316899 0.31170295 0.4056147
#> [10,] 0.005959664 0.23267892 0.34692799 0.32659695 0.26795300 0.2123354
#> [11,] 0.034529005 0.36358010 0.48645503 0.50762936 0.59931211 0.5886570
#> [12,] 0.034219221 0.45080291 0.53811127 0.52949794 0.48895181 0.3641406
#> [13,] 0.020144284 0.30317561 0.26487055 0.28695418 0.24412975 0.1929554
#> [14,] 0.022393410 0.23214817 0.27365070 0.32232154 0.28050113 0.2165524
#> [15,] 0.015508070 0.11155357 0.19711484 0.33924844 0.31373585 0.2557876
#> [16,] 0.015499675 0.07601012 0.09344890 0.11879800 0.17324102 0.1488400
#> [17,] 0.029807753 0.10047009 0.08554889 0.07925629 0.01531301 0.1091627
#> [18,] 0.019383989 0.13340329 0.21481753 0.16107887 0.14709679 0.1802539
#> [19,] 0.030735092 0.04656500 0.10423448 0.21160371 0.17224192 0.1273992
#> [20,] 0.015892145 0.12277221 0.15576330 0.22425870 0.27970815 0.2666823
#> [21,] 0.018199528 0.15410343 0.13842948 0.09151329 0.11211347 0.1591787
#> [22,] 0.019029703 0.09091432 0.11607286 0.11199465 0.12100668 0.1138122
#> [23,] 0.027175577 0.07610939 0.03649766 0.20918112 0.18788780 0.1078577
#> [24,] 0.019771419 0.08153622 0.09667009 0.14475614 0.23972319 0.2577701
#> [25,] 0.040489285 0.34201682 0.27649241 0.06864391 0.20229347 0.4920220
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.38401326 0.4285521 0.4433458 0.4489867 0.4882561 0.194365743 0.3514219
#>  [2,] 0.48016567 0.4643030 0.5618683 0.6599174 0.7159156 0.016466745 0.7607447
#>  [3,] 0.23969339 0.2169818 0.2540017 0.3792837 0.5659869 0.054819434 0.8288238
#>  [4,] 0.15446768 0.1773152 0.2083574 0.2514312 0.2709814 0.024501011 0.3101857
#>  [5,] 0.34756349 0.4271102 0.4695448 0.3426323 0.4531930 0.087593831 0.5487986
#>  [6,] 0.22454812 0.2079445 0.1945305 0.1572684 0.2748419 0.033228811 0.3394232
#>  [7,] 0.30468986 0.3208922 0.2857625 0.3035903 0.2668867 0.033214358 0.5667576
#>  [8,] 0.05801493 0.1469006 0.2859135 0.3919802 0.5062189 0.016275173 0.5785191
#>  [9,] 0.47531854 0.5821319 0.6832667 0.7442886 0.8214738 0.045545968 0.8414487
#> [10,] 0.13798906 0.1148191 0.1617143 0.2534678 0.3340801 0.025611879 0.3548950
#> [11,] 0.66700056 0.7249823 0.8633866 0.9668487 0.9128822 0.035416110 0.8355617
#> [12,] 0.26336614 0.2774982 0.3064422 0.3310039 0.4919998 0.044808959 0.6706056
#> [13,] 0.13214531 0.1725088 0.2001127 0.2359419 0.1873512 0.031292238 0.1183726
#> [14,] 0.16572061 0.2105654 0.2272243 0.3366671 0.5327458 0.027023384 0.6502503
#> [15,] 0.24508751 0.4784197 0.6898038 0.6708430 0.5205387 0.034672243 0.4316672
#> [16,] 0.17667562 0.2614904 0.1887182 0.2687330 0.4348701 0.040262046 0.1890451
#> [17,] 0.09471361 0.1986123 0.1567101 0.2737675 0.3823735 0.039241950 0.3838475
#> [18,] 0.18022266 0.1946332 0.2472496 0.1071088 0.3939944 0.002274732 0.5995582
#> [19,] 0.12116299 0.1726157 0.2309380 0.1559171 0.2020575 0.040191222 0.2133561
#> [20,] 0.29237091 0.3413001 0.3739056 0.3191937 0.2358558 0.017548015 0.2811929
#> [21,] 0.14308868 0.1569620 0.2200008 0.2346666 0.2825055 0.011498776 0.4074174
#> [22,] 0.13018745 0.1160823 0.1462724 0.2669555 0.3199226 0.020631845 0.3903164
#> [23,] 0.14416519 0.2195073 0.2027629 0.1116806 0.1129095 0.043757317 0.1163098
#> [24,] 0.27601792 0.2990364 0.2709489 0.2871203 0.4087459 0.020758908 0.5320886
#> [25,] 0.62101593 0.7132634 0.7859838 0.6734249 0.4450803 0.098359434 0.3752936
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.2274363 0.17984385 0.3969513 0.6250762 0.48036183 0.3061350 0.4408542
#>  [2,] 0.8265556 0.90641505 0.9172980 0.9595630 0.95386799 0.9533872 0.9341083
#>  [3,] 0.8174815 0.40401593 0.2027089 0.2069105 0.34934893 0.6689583 0.7338821
#>  [4,] 0.4151862 0.50485628 0.5877852 0.6928581 0.80608485 0.8724434 0.9542336
#>  [5,] 0.5857435 0.64109675 0.6268386 0.6290515 0.63003462 0.6335068 0.6762344
#>  [6,] 0.4269602 0.47465877 0.4709124 0.4561097 0.46633610 0.5707787 0.7569188
#>  [7,] 0.6982580 0.65593856 0.5936913 0.6825542 0.62208347 0.8449056 0.9922949
#>  [8,] 0.6694053 0.74776785 0.7450737 0.7173938 0.78931478 0.9138449 0.9780221
#>  [9,] 0.8870119 0.92965519 0.9165019 0.7791314 0.59052362 0.5265603 0.4712036
#> [10,] 0.3197875 0.33913737 0.3548528 0.4646284 0.60377406 0.7878105 0.9313209
#> [11,] 0.6924372 0.42727968 0.1786812 0.1232382 0.26723650 0.3488448 0.3169833
#> [12,] 0.7835791 0.87872491 0.8456000 0.8532277 0.83788516 0.8495582 0.8782477
#> [13,] 0.0815082 0.21612291 0.4253821 0.5774060 0.69769985 0.8280789 0.9281583
#> [14,] 0.7726409 0.78654195 0.7959977 0.9042910 0.91811958 0.9517428 0.9819316
#> [15,] 0.4075617 0.40912240 0.5336478 0.7655017 0.86552769 0.8106151 0.7978277
#> [16,] 0.3724021 0.41248193 0.3538711 0.4785169 0.43215531 0.5040755 0.3460123
#> [17,] 0.3243339 0.13288592 0.1181558 0.2861238 0.15220948 0.2596648 0.4522429
#> [18,] 0.3770127 0.38194503 1.0233263 0.8905574 0.43468749 0.5274944 0.6396046
#> [19,] 0.1561146 0.11736129 0.1665688 0.1140028 0.16926636 0.3433496 0.4971191
#> [20,] 0.4467288 0.48394799 0.4846453 0.6328075 0.85071894 0.8798720 0.7844884
#> [21,] 0.5185325 0.68239663 0.8543083 0.8787091 0.84047676 0.8880183 0.9143230
#> [22,] 0.4902469 0.71943879 0.8249711 0.8231041 0.78464664 0.7810929 0.6810406
#> [23,] 0.2213834 0.06581727 0.2082301 0.2394691 0.08604482 0.2254716 0.5798605
#> [24,] 0.5217344 0.58008323 0.6252695 0.6377330 0.79209574 0.9086780 0.9816708
#> [25,] 0.5548821 0.74750357 0.8338285 0.9052281 0.82144427 0.6909991 0.6748728
#>             V28       V29         V3       V30       V31       V32       V33
#>  [1,] 0.5777915 0.5097976 0.25421247 0.4194507 0.3745982 0.2697712 0.1646670
#>  [2,] 0.8572301 0.7870035 0.03372060 0.6403217 0.4220440 0.2907967 0.1347389
#>  [3,] 0.7731120 0.7553764 0.05454743 0.7663507 0.4486035 0.4376136 0.6681114
#>  [4,] 0.9851769 0.8508404 0.03422916 0.6834161 0.6583681 0.5189726 0.3420325
#>  [5,] 0.7489287 0.7249189 0.09979832 0.8135120 0.8338140 0.7831331 0.5709600
#>  [6,] 0.8693989 0.9688087 0.02596195 0.9251880 0.6788647 0.5170590 0.4666075
#>  [7,] 0.9288232 0.8110797 0.07053750 0.6681986 0.5664759 0.7207944 0.5894879
#>  [8,] 0.9679675 0.8478159 0.03229641 0.6072109 0.2424038 0.2115095 0.3144619
#>  [9,] 0.4311039 0.4405595 0.03872878 0.3935348 0.3047059 0.3356953 0.3520049
#> [10,] 1.0506108 0.9745791 0.02524144 0.6652455 0.4933307 0.5316234 0.4986928
#> [11,] 0.2649198 0.1100808 0.07689457 0.1873640 0.2757304 0.3301406 0.3501334
#> [12,] 0.7925306 0.6283107 0.04052542 0.4555594 0.4064910 0.4293800 0.4611345
#> [13,] 0.9434468 0.9056979 0.03273427 0.9369079 0.9135879 0.8228794 0.7153443
#> [14,] 0.9646033 0.8486617 0.03857211 0.6341188 0.4317982 0.3068356 0.2504278
#> [15,] 0.8573173 0.8992303 0.03388318 0.7852224 0.7165680 0.6898556 0.7069077
#> [16,] 0.3847643 0.6339017 0.03745016 0.7679080 0.7353406 0.5026259 0.4540628
#> [17,] 0.7577914 0.9196664 0.07578868 0.7829173 0.4866487 0.2471971 0.5341822
#> [18,] 0.7538024 0.5898261 0.05505345 0.3985618 0.2628661 0.1906597 0.1072431
#> [19,] 0.6219172 0.6754535 0.04623810 0.6829705 0.8042449 0.7117088 0.6915474
#> [20,] 0.7427501 0.6920236 0.01766564 0.5010851 0.3621931 0.2530945 0.1736841
#> [21,] 0.9069512 0.9094177 0.01903858 0.7911328 0.6909149 0.6178939 0.5830264
#> [22,] 0.6668989 0.6890021 0.03679632 0.6217986 0.5372552 0.5973601 0.7392377
#> [23,] 0.8361632 0.8359301 0.01488062 0.7597931 0.7989935 0.6201822 0.6548673
#> [24,] 0.9286493 0.7456287 0.01602536 0.6346196 0.5742086 0.4093125 0.2904506
#> [25,] 0.5543979 0.3840201 0.11028158 0.4355486 0.4588536 0.2518219 0.2077637
#>              V34        V35        V36        V37       V38        V39
#>  [1,] 0.19604781 0.27511511 0.40073023 0.31674762 0.3111189 0.36895144
#>  [2,] 0.02411924 0.04484390 0.16782088 0.22630472 0.2451061 0.26891272
#>  [3,] 0.71536815 0.63304694 0.53179637 0.51152247 0.5651037 0.49009274
#>  [4,] 0.22545143 0.13173125 0.08171003 0.18633025 0.1950785 0.18609917
#>  [5,] 0.49234136 0.47617921 0.24526021 0.22968505 0.3533984 0.35537631
#>  [6,] 0.36437029 0.19557424 0.17973775 0.24402369 0.1975450 0.17820977
#>  [7,] 0.25991668 0.03424213 0.20080429 0.30561853 0.4675819 0.58264865
#>  [8,] 0.50263027 0.74157651 0.86577483 0.79501241 0.6532530 0.48215611
#>  [9,] 0.32712485 0.35906490 0.38389559 0.35549823 0.3489166 0.34704474
#> [10,] 0.33116040 0.15287148 0.11674902 0.17202749 0.1035005 0.08756225
#> [11,] 0.35384519 0.41602991 0.09676581 0.16422999 0.2788614 0.20471689
#> [12,] 0.38884370 0.20398711 0.31410323 0.34081614 0.3461012 0.44088993
#> [13,] 0.65062140 0.66060713 0.70174075 0.67486920 0.5484468 0.45056310
#> [14,] 0.18345144 0.13263512 0.04808580 0.07181012 0.1113575 0.10654344
#> [15,] 0.85850491 0.93474315 0.93827626 0.90073606 0.7415510 0.57663251
#> [16,] 0.50595118 0.60716707 0.66875810 0.65186827 0.7826815 0.93927771
#> [17,] 0.61927044 0.34683013 0.43992651 0.62882966 0.6366005 0.54768522
#> [18,] 0.30421293 0.26336656 0.34202660 0.27249769 0.1932686 0.23519350
#> [19,] 0.81013248 0.92502714 0.97835331 0.85840168 0.7466054 0.64546479
#> [20,] 0.16217278 0.22875401 0.19625360 0.13325720 0.1517252 0.14192875
#> [21,] 0.60659516 0.60906826 0.54129252 0.44225682 0.3290943 0.21365913
#> [22,] 0.81751687 0.78890239 0.65959869 0.45045211 0.2479295 0.21957590
#> [23,] 0.87444420 0.98609285 0.93648566 0.70093237 0.5801148 0.59086648
#> [24,] 0.16859613 0.22027445 0.36676352 0.31616794 0.2656428 0.30255118
#> [25,] 0.13496052 0.44118974 0.66753254 0.66260924 0.3800856 0.19312734
#>                V4       V40         V41         V42          V43          V44
#>  [1,] 0.354573930 0.4081133 0.346042082  0.27726949  0.184076709  0.194604790
#>  [2,] 0.084043261 0.2974673 0.277498883  0.32638994  0.275715195  0.157060813
#>  [3,] 0.097185481 0.3769913 0.407158097  0.40531309  0.342523205  0.234704395
#>  [4,] 0.043481724 0.1727557 0.214132225  0.21019341  0.147998059  0.126409370
#>  [5,] 0.110496141 0.2057101 0.544645683  0.61392468  0.422252520  0.386454303
#>  [6,] 0.068534206 0.3368241 0.460109344  0.47158171  0.405266291  0.208358032
#>  [7,] 0.018587082 0.4351070 0.471934533  0.68457666  0.598080913  0.556337905
#>  [8,] 0.042807816 0.3378485 0.305024018  0.28539329  0.350587197  0.386777448
#>  [9,] 0.057470806 0.2985256 0.261665107  0.32264552  0.318769866  0.271688342
#> [10,] 0.022562171 0.2548652 0.340043636  0.49432720  0.424014446  0.119249664
#> [11,] 0.104522545 0.1384522 0.144428575  0.19229160  0.132514696  0.081165037
#> [12,] 0.036996963 0.4021139 0.205770626  0.28392122  0.367754651  0.392632753
#> [13,] 0.052679821 0.4961880 0.476038309  0.42011492  0.447709263  0.509051680
#> [14,] 0.038064600 0.1145440 0.058368051  0.12122824  0.207613583  0.225829524
#> [15,] 0.044486485 0.4350334 0.291407363  0.20949598  0.134057691  0.122204940
#> [16,] 0.018969563 0.9353586 0.804034898  0.74126870  0.696761543  0.630915026
#> [17,] 0.079484494 0.3210499 0.218885488  0.05396158 -0.005593764 -0.005205757
#> [18,] 0.051390302 0.1017235 0.005596529 -0.01640817  0.052143502  0.093921399
#> [19,] 0.011917138 0.5813636 0.486464768  0.44269926  0.360410096  0.260858587
#> [20,] 0.027646947 0.1213168 0.185632386  0.18570491  0.184634806  0.146396969
#> [21,] 0.028103103 0.1673627 0.159016357  0.10549175  0.146449277  0.098251042
#> [22,] 0.036215062 0.3531174 0.376958177  0.36479948  0.402377508  0.290447037
#> [23,] 0.009770831 0.6666739 0.567618656  0.52409657  0.369173962  0.201543079
#> [24,] 0.020260287 0.2181782 0.141530658  0.23398120  0.180505013  0.146822032
#> [25,] 0.095260329 0.3685367 0.396388583  0.34361221  0.243068399  0.203414991
#>                V45          V46          V47          V48          V49
#>  [1,]  0.265645324  0.214386296  0.169815554  0.140405308  0.066004790
#>  [2,]  0.099581900  0.125736192  0.146929926  0.102257312  0.054389534
#>  [3,]  0.205110825  0.170033112  0.090226367  0.106656406  0.052120313
#>  [4,]  0.119812043  0.149201088  0.158758506  0.117044730  0.052339239
#>  [5,]  0.544716494  0.535600396  0.313209399  0.170678831  0.107053237
#>  [6,]  0.282573787  0.267695016  0.232610239  0.193951579  0.120345809
#>  [7,]  0.568791168  0.340812393  0.237495746  0.224381913  0.122749975
#>  [8,]  0.322159319  0.216313302  0.102519237  0.049286598  0.036051779
#>  [9,]  0.153641452  0.107659096  0.103242038  0.069032600  0.055742446
#> [10,]  0.173297851  0.124184379  0.175189306  0.168543201  0.085353078
#> [11,]  0.070344344  0.048241291  0.049090689  0.046322864  0.024761806
#> [12,]  0.373791907  0.181823802  0.102950493  0.110556933  0.063831436
#> [13,]  0.457170091  0.329678416  0.232762237  0.182717025  0.100731242
#> [14,]  0.206351744  0.209409120  0.172919483  0.109092688  0.052764298
#> [15,]  0.172051713  0.187240347  0.113315497  0.063262307  0.022985624
#> [16,]  0.560482404  0.475274847  0.342902793  0.208129789  0.117438938
#> [17,] -0.009563647 -0.006488335 -0.006103921 -0.009193604 -0.004056107
#> [18,]  0.047925370  0.005670567  0.011095486 -0.001352101  0.005912175
#> [19,]  0.159448920  0.089983067  0.093189684  0.085538192  0.051391219
#> [20,]  0.088333043  0.067099323  0.045412239  0.036436443  0.019634126
#> [21,]  0.036547855  0.041904609  0.035723319  0.039975356  0.012787824
#> [22,]  0.155820889  0.085190503  0.061144557  0.046888053  0.018491708
#> [23,]  0.151265590  0.052721787  0.092267007  0.096691430  0.052675478
#> [24,]  0.132567053  0.085491567  0.086944279  0.070561783  0.023305899
#> [25,]  0.210714632  0.161359488  0.155355485  0.153794122  0.090200199
#>               V5           V50          V51         V52          V53
#>  [1,] 0.33105160  0.0184746502  0.014052469 0.032671629 0.0204222551
#>  [2,] 0.13551473  0.0272787076  0.016166283 0.010095676 0.0108555646
#>  [3,] 0.13900667  0.0212835605  0.020341717 0.017270339 0.0181260529
#>  [4,] 0.06662940  0.0186799457  0.017643888 0.012926431 0.0076685581
#>  [5,] 0.12852362  0.0499712551  0.035732802 0.026430468 0.0212981457
#>  [6,] 0.08629578  0.0193924764  0.015328797 0.014603791 0.0008814468
#>  [7,] 0.10454304  0.0326194272  0.035934718 0.018418986 0.0040773182
#>  [8,] 0.04278559  0.0131942257  0.004848986 0.009040164 0.0078663862
#>  [9,] 0.08967227  0.0174039655  0.015910179 0.010382239 0.0105426723
#> [10,] 0.01957799  0.0195108607  0.022508290 0.008831919 0.0113461509
#> [11,] 0.05057532  0.0166891308  0.026646451 0.022605921 0.0240004089
#> [12,] 0.04406171  0.0145111855  0.019709035 0.011760126 0.0073507866
#> [13,] 0.07857458  0.0301245929  0.016369624 0.013233019 0.0086106774
#> [14,] 0.07039701  0.0177387002  0.017565603 0.012152548 0.0062526503
#> [15,] 0.05315495  0.0209312189  0.023180708 0.015395572 0.0094413325
#> [16,] 0.09998293  0.0348739192  0.025386320 0.011929186 0.0139024840
#> [17,] 0.05252393 -0.0007221236 -0.001724474 0.004960507 0.0239227446
#> [18,] 0.06571029  0.0087372401  0.033525862 0.045683368 0.0160698470
#> [19,] 0.01736510  0.0145250105  0.005494942 0.008905258 0.0201646004
#> [20,] 0.04242457  0.0107457672  0.007562191 0.007847176 0.0068071284
#> [21,] 0.04894517  0.0101189865  0.006844399 0.006046474 0.0099434514
#> [22,] 0.04654934  0.0140402178  0.012686674 0.010531536 0.0144728779
#> [23,] 0.03387789  0.0187147330  0.009724522 0.004147627 0.0185229070
#> [24,] 0.03884343  0.0127975403  0.007049607 0.006124627 0.0087445386
#> [25,] 0.09649353  0.0198436815  0.018819614 0.013914849 0.0101988061
#>                V54           V55         V56          V57         V58
#>  [1,]  0.015523802  0.0202697121 0.002888798 0.0134061895 0.014701473
#>  [2,]  0.019087878  0.0053034093 0.012578812 0.0079039350 0.019008640
#>  [3,]  0.017746406  0.0071463052 0.011220489 0.0066168818 0.007109505
#>  [4,]  0.006574130  0.0051991523 0.005410233 0.0051427301 0.005825030
#>  [5,]  0.019766617  0.0205990983 0.016884720 0.0174435527 0.019380347
#>  [6,]  0.002858650  0.0033252718 0.010034762 0.0049556333 0.008884493
#>  [7,]  0.009984726  0.0127578742 0.006654162 0.0063115782 0.002301390
#>  [8,]  0.003860086  0.0031829123 0.004290274 0.0023339821 0.001485520
#>  [9,]  0.015131423  0.0087674616 0.008896441 0.0075063519 0.013962619
#> [10,]  0.010339248 -0.0005432068 0.002804629 0.0001674673 0.007944508
#> [11,]  0.014113358  0.0197498650 0.016108615 0.0051844307 0.028349856
#> [12,]  0.011357450  0.0097263335 0.006849433 0.0055542430 0.008475807
#> [13,]  0.006421077  0.0056018900 0.007796020 0.0056411469 0.004964697
#> [14,]  0.005777600  0.0040070274 0.003684912 0.0036455103 0.002187045
#> [15,]  0.007657199  0.0114628594 0.007576458 0.0092619211 0.009055228
#> [16,]  0.010540729  0.0144318847 0.017225525 0.0015474765 0.017181564
#> [17,]  0.019868287  0.0046119933 0.013020988 0.0162670718 0.007787439
#> [18,] -0.002131450  0.0097298719 0.009429564 0.0085585168 0.019615698
#> [19,]  0.010186209  0.0103701753 0.004262362 0.0060129003 0.014467653
#> [20,]  0.006979016  0.0108014818 0.007087327 0.0058536721 0.005192611
#> [21,]  0.007692942  0.0050714625 0.004986698 0.0053236090 0.005300448
#> [22,]  0.012313387  0.0049164347 0.012292564 0.0152302633 0.005953117
#> [23,]  0.015626236  0.0068852670 0.009522907 0.0077066581 0.008008174
#> [24,]  0.010918313  0.0055785707 0.009222730 0.0066657448 0.005471526
#> [25,]  0.015189876  0.0110593893 0.006359235 0.0115897968 0.009848341
#>               V59         V6          V60          V7         V8         V9
#>  [1,] 0.018047605 0.16011695 0.0203144443 0.171846798 0.06022069 0.17798180
#>  [2,] 0.009215658 0.16538269 0.0079547005 0.156903570 0.16940424 0.23838199
#>  [3,] 0.009029735 0.07224320 0.0058847932 0.087720502 0.18950559 0.42445830
#>  [4,] 0.002604662 0.10208199 0.0038872899 0.116282993 0.11341616 0.16124077
#>  [5,] 0.017737683 0.09873336 0.0130691118 0.099794383 0.11163155 0.20763802
#>  [6,] 0.009083978 0.08214374 0.0055722711 0.074813380 0.12591428 0.20238566
#>  [7,] 0.004300000 0.09674953 0.0017134260 0.096019444 0.14326904 0.24732778
#>  [8,] 0.003280955 0.05063522 0.0033821312 0.057544663 0.03468522 0.10206246
#>  [9,] 0.012213076 0.10366904 0.0077253986 0.126577663 0.10523041 0.14048944
#> [10,] 0.006112628 0.07790838 0.0012850655 0.138678438 0.11044635 0.16040447
#> [11,] 0.012656609 0.05751033 0.0134966546 0.121933345 0.07099524 0.19213395
#> [12,] 0.007120338 0.09925451 0.0052724854 0.113734719 0.19228205 0.32882731
#> [13,] 0.003607663 0.10762519 0.0036798000 0.133685029 0.15705105 0.27301116
#> [14,] 0.002308303 0.08498076 0.0044836963 0.109304233 0.13006381 0.18103673
#> [15,] 0.008959245 0.12614774 0.0097216575 0.156665985 0.10512047 0.08744992
#> [16,] 0.018879219 0.10941810 0.0148648607 0.166201969 0.21350075 0.04905183
#> [17,] 0.013519179 0.10446769 0.0109455158 0.006057261 0.05341739 0.11038408
#> [18,] 0.011826043 0.16808776 0.0067727447 0.145740310 0.22040633 0.14500573
#> [19,] 0.016908606 0.04200202 0.0005074015 0.038946673 0.05254659 0.03000106
#> [20,] 0.003755080 0.06776899 0.0030945084 0.093524284 0.08233984 0.09787348
#> [21,] 0.005506604 0.08673568 0.0042222967 0.078716249 0.08652657 0.14609354
#> [22,] 0.007217117 0.07031244 0.0104041978 0.087723695 0.08338084 0.03337003
#> [23,] 0.008547528 0.04972934 0.0093229070 0.039478381 0.07723260 0.07988085
#> [24,] 0.006091922 0.06334744 0.0064607494 0.087215736 0.08653836 0.08201152
#> [25,] 0.009689363 0.12631470 0.0078193910 0.180071197 0.22271681 0.27569016
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551