Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.042294126  0.26367782 0.29983356 0.34035037 0.32976627 0.25745331
#>  [2,] 0.031534870  0.50653240 0.33849498 0.10533226 0.27689753 0.37366643
#>  [3,] 0.044659566  0.06569532 0.14600518 0.19892026 0.24587660 0.30778587
#>  [4,] 0.030132744  0.13511738 0.25619911 0.36741970 0.35191551 0.38801977
#>  [5,] 0.043212986  0.30248444 0.39084338 0.54160966 0.58929162 0.71740100
#>  [6,] 0.043959856  0.20234884 0.24028627 0.34877606 0.41791970 0.50270421
#>  [7,] 0.016294040  0.25546018 0.27046095 0.21619220 0.36480885 0.58294811
#>  [8,] 0.062900000  0.35970000 0.54660000 0.52050000 0.51270000 0.53950000
#>  [9,] 0.027732530  0.26817487 0.29485384 0.31747327 0.32939209 0.37976051
#> [10,] 0.016963754  0.24977178 0.29046480 0.34783073 0.29773425 0.22319444
#> [11,] 0.102464075  0.60646938 0.51566123 0.23347916 0.20414794 0.30560749
#> [12,] 0.043695735  0.25374548 0.35455690 0.36839882 0.37906290 0.35940214
#> [13,] 0.034120517  0.45408059 0.58552889 0.59636135 0.71079859 0.69237746
#> [14,] 0.009750852  0.07704812 0.06777799 0.03823062 0.06452461 0.07571558
#> [15,] 0.015435719  0.02529327 0.05876834 0.05298350 0.10215224 0.10619151
#> [16,] 0.023744799  0.01937417 0.03127729 0.11178965 0.12020671 0.11283322
#> [17,] 0.009393247 -0.05715144 0.07356784 0.36059494 0.31457148 0.30720893
#> [18,] 0.018484747  0.06338867 0.07707357 0.10574170 0.19762059 0.36098402
#> [19,] 0.035448775  0.02850579 0.09631304 0.24182794 0.18905992 0.10059388
#> [20,] 0.058141021  0.31090185 0.19523247 0.23561241 0.15945599 0.10207239
#> [21,] 0.029612298  0.32550443 0.26214402 0.10497037 0.22892823 0.49620185
#> [22,] 0.006589387  0.08564783 0.07346720 0.02462023 0.08880760 0.15247511
#> [23,] 0.019113034  0.13457788 0.11557946 0.06106637 0.05710011 0.12572293
#> [24,] 0.026810464  0.11042752 0.08946319 0.07549199 0.01611543 0.11422300
#> [25,] 0.006817262  0.08042376 0.03477207 0.05207984 0.24626206 0.25341345
#>                V15         V16       V17        V18       V19           V2
#>  [1,]  0.199540131  0.26701965 0.2918718 0.26833069 0.3936113  0.072531349
#>  [2,]  0.388423302  0.42038369 0.1548125 0.14949481 0.2482990  0.066418447
#>  [3,]  0.386925530  0.42478229 0.5598033 0.68254390 0.8554116  0.062228929
#>  [4,]  0.533806868  0.61660852 0.7505582 0.80474903 0.8773749  0.034422832
#>  [5,]  0.759287384  0.85343561 0.9673091 0.89544683 0.9200876  0.046194321
#>  [6,]  0.585721435  0.63568140 0.7102503 0.80605926 0.9025830  0.044806161
#>  [7,]  0.657349679  0.72714545 0.6961881 0.78116188 0.9068848  0.029387342
#>  [8,]  0.655800000  0.87050000 0.9786000 0.93350000 0.7917000  0.106500000
#>  [9,]  0.422338637  0.50896044 0.6263549 0.69548311 0.7601525  0.038095099
#> [10,]  0.163167856  0.18902893 0.2096609 0.32878578 0.5792249  0.028199440
#> [11,]  0.592336788  0.59190489 0.1976166 0.14769905 0.4368936  0.102825379
#> [12,]  0.450799175  0.49666013 0.6081622 0.70677085 0.8187339  0.049621470
#> [13,]  0.703356713  0.70621260 0.8509800 0.95128334 0.8978033  0.009360308
#> [14,] -0.004950782  0.09485140 0.1562379 0.42777967 0.3463457  0.004257171
#> [15,]  0.106322007  0.19247751 0.3123621 0.36879623 0.3927762  0.019272401
#> [16,]  0.140601397  0.13817974 0.1328014 0.01477983 0.1155443  0.038987069
#> [17,] -0.030829021 -0.05908614 0.1268369 0.46380914 0.2492820 -0.009670616
#> [18,]  0.502677641  0.66381070 0.7354515 0.75884992 0.7825169  0.022543061
#> [19,]  0.143181993  0.20274744 0.2059471 0.20856827 0.2761064  0.049295898
#> [20,]  0.158654729  0.26683459 0.4066079 0.64888639 0.7500637  0.061891892
#> [21,]  0.643953428  0.80324438 0.8509289 0.76370045 0.5248033  0.059764313
#> [22,]  0.211905288  0.20657939 0.1230268 0.13865367 0.1670873 -0.001595171
#> [23,]  0.102989300  0.11893441 0.2279024 0.22600558 0.2748008  0.004118199
#> [24,]  0.099620803  0.20255875 0.1509318 0.26835380 0.3783040  0.036251130
#> [25,]  0.155804012  0.20140291 0.1462291 0.31291957 0.6867201  0.022405652
#>             V20       V21        V22         V23          V24         V25
#>  [1,] 0.4701097 0.5417314 0.63306612  0.66156559  0.647803815  0.62382480
#>  [2,] 0.8968424 0.9834970 0.34487379  0.16947928  0.188767601  0.12734999
#>  [3,] 0.9314195 0.9226329 0.99235786  0.97646573  0.939944637  0.87342010
#>  [4,] 0.9119805 0.8763428 0.87023722  0.85495439  0.784501544  0.54140678
#>  [5,] 0.8657690 0.7999711 0.77748435  0.63441912  0.361040982  0.23139925
#>  [6,] 0.9034518 0.9827893 0.90384014  0.79105423  0.837629586  0.74890751
#>  [7,] 0.9751059 1.0248193 1.02804047  1.01103490  0.876659895  0.70251536
#>  [8,] 0.7383000 0.6908000 0.38500000  0.06710000  0.050200000  0.27170000
#>  [9,] 0.7838323 0.7802442 0.83850945  0.92177675  0.943740794  0.91763020
#> [10,] 0.7104181 0.8311332 0.84264704  0.83668462  0.936424728  0.93131843
#> [11,] 0.8820672 0.8584693 0.38805734  0.15095287  0.133493836  0.22313772
#> [12,] 0.8329735 0.9230588 0.95614322  0.91952619  0.845695080  0.71588458
#> [13,] 0.8266441 0.6226781 0.29439882 -0.01115905 -0.009443287  0.26187111
#> [14,] 0.2995512 0.1928869 0.18776049  0.16566004  0.181590461 -0.01265688
#> [15,] 0.3854453 0.4949342 0.58946066  0.62057946  0.685137917  0.68404439
#> [16,] 0.2325240 0.2044310 0.20258237  0.20179180  0.103179368  0.12073228
#> [17,] 0.5415095 0.6849120 0.70578357  0.77300418  0.859311215  1.08620945
#> [18,] 0.8444967 0.9515049 0.90602346  0.86525262  0.907032562  0.90836732
#> [19,] 0.1593496 0.2538935 0.06029982  0.20922182  0.186287730 -0.01322232
#> [20,] 0.8402875 0.8392895 0.80797679  0.89512496  0.749758036  0.37354690
#> [21,] 0.4028643 0.5409021 0.66783398  0.79226757  0.855272868  0.79323228
#> [22,] 0.1959077 0.2141366 0.30632319  0.51518013  0.686765493  0.80842285
#> [23,] 0.4527242 0.5864816 0.76363871  0.94078819  0.926151510  0.88274145
#> [24,] 0.4047697 0.3505800 0.13499819  0.11031784  0.274501723  0.13940161
#> [25,] 0.8920286 0.7924290 0.74621839  0.49688181  0.392888455  0.59652896
#>             V26       V27       V28       V29           V3       V30        V31
#>  [1,] 0.6853331 0.7963020 0.8200284 0.7406000  0.077296652 0.6908078 0.64304341
#>  [2,] 0.3530881 0.2937804 0.5285742 0.4364685  0.100759749 0.6796837 0.20568810
#>  [3,] 0.7522176 0.6977787 0.5331391 0.3581840  0.146290200 0.2884238 0.18416693
#>  [4,] 0.3863348 0.3037919 0.3583204 0.3863336  0.048964521 0.4133567 0.32807462
#>  [5,] 0.1699874 0.2554786 0.2180526 0.1569174  0.032107072 0.1475349 0.08442232
#>  [6,] 0.6506523 0.5644880 0.3615766 0.3361397  0.043605478 0.3325220 0.19768736
#>  [7,] 0.7232306 0.6362029 0.5422969 0.4824195  0.032984526 0.3810760 0.31865991
#>  [8,] 0.2839000 0.2234000 0.1911000 0.0408000  0.152600000 0.2531000 0.19790000
#>  [9,] 0.9641020 0.9712656 0.8222465 0.6488236  0.049687243 0.5268524 0.36137230
#> [10,] 0.9740077 0.9964492 0.9773112 0.8710806  0.041984079 0.6401446 0.41543178
#> [11,] 0.3493995 0.2439170 0.5955794 0.5976097  0.037746121 0.7440186 0.42005906
#> [12,] 0.5770036 0.3938543 0.2891326 0.3141750  0.068350152 0.4039508 0.38332420
#> [13,] 0.4745495 0.4981158 0.4484567 0.2775577  0.074906053 0.1382900 0.33041951
#> [14,] 0.1485588 0.4064065 0.6407302 0.7929291 -0.021279387 0.7205598 0.49769660
#> [15,] 0.7187706 0.7060963 0.7616748 0.8718513  0.023132408 0.7840017 0.66252810
#> [16,] 0.2945073 0.4511601 0.5684008 0.6012723  0.039433367 0.6486275 0.71705410
#> [17,] 0.9284691 0.6830392 0.4953751 0.5344128  0.065862025 0.4571157 0.43485103
#> [18,] 0.9220965 0.8777040 0.6807518 0.4169318  0.009469837 0.3538159 0.34428416
#> [19,] 0.1230934 0.2841666 0.5801636 0.7004535  0.038339034 0.6786070 0.69562626
#> [20,] 0.1429127 0.2911057 0.4258254 0.4376575  0.075467842 0.5415813 0.73456319
#> [21,] 0.7009872 0.6944719 0.5705453 0.3247819  0.077943973 0.3660495 0.50150006
#> [22,] 0.8858918 0.7717339 0.7571912 0.7320172  0.013828188 0.6051326 0.43950592
#> [23,] 0.9419996 0.9648511 0.9072893 0.8789095  0.009208885 0.7669768 0.70129714
#> [24,] 0.2606649 0.4483998 0.7567280 0.9109738  0.072713463 0.7849048 0.47236427
#> [25,] 0.8059244 0.9571646 0.8134752 0.5675427  0.012905445 0.3944014 0.38385263
#>              V32       V33       V34       V35        V36        V37        V38
#>  [1,] 0.58319174 0.4440165 0.4000896 0.3264792 0.34350236 0.36865966 0.36233752
#>  [2,] 0.29280264 0.5849586 0.8031390 0.8135755 0.88722149 0.80810851 0.90524982
#>  [3,] 0.10625380 0.1491524 0.2427436 0.4274047 0.46861588 0.39541808 0.37973708
#>  [4,] 0.40398501 0.3699435 0.2061775 0.1417408 0.19901798 0.31366305 0.45586124
#>  [5,] 0.12953650 0.2237442 0.1621925 0.2141413 0.12866293 0.28592629 0.28219941
#>  [6,] 0.07638736 0.3272538 0.4544315 0.4264902 0.38560782 0.30548981 0.34087795
#>  [7,] 0.22163536 0.1133804 0.1293796 0.2082224 0.27916117 0.22901253 0.17228235
#>  [8,] 0.18910000 0.2433000 0.1956000 0.2667000 0.13400000 0.10730000 0.20230000
#>  [9,] 0.30571590 0.2088025 0.1441773 0.1648791 0.15254424 0.21344819 0.27294033
#> [10,] 0.31230813 0.2518572 0.1994706 0.1676920 0.04909547 0.04928809 0.07625111
#> [11,] 0.43622180 0.6970848 0.8198881 0.9361440 0.86787717 0.87696314 0.90446792
#> [12,] 0.23829239 0.3855406 0.5971769 0.6268980 0.52372960 0.40958714 0.32272096
#> [13,] 0.33629451 0.3323786 0.3227745 0.3920525 0.04878109 0.15841470 0.15239248
#> [14,] 0.12127250 0.4648514 0.4692132 0.2150874 0.35426719 0.49237231 0.59892692
#> [15,] 0.63788870 0.7036175 0.7381537 0.7240797 0.66287862 0.55249673 0.46179923
#> [16,] 0.60437326 0.6986508 0.8948388 1.0053899 1.01826640 0.78168571 0.60753178
#> [17,] 0.42922932 0.5799950 0.4929160 0.5113598 0.54917015 0.50136594 0.24720647
#> [18,] 0.28928800 0.2658165 0.2132287 0.1536491 0.09096573 0.13843008 0.14050653
#> [19,] 0.53866774 0.6414548 0.8373970 0.9079565 0.93457270 0.79749761 0.78915901
#> [20,] 0.71296648 0.5876509 0.5763939 0.5571135 0.65070704 0.64168031 0.75222063
#> [21,] 0.43037921 0.2559586 0.2515950 0.5781566 0.67649310 0.59357933 0.46150573
#> [22,] 0.31031839 0.3089157 0.4527792 0.4898240 0.41808169 0.15359744 0.09293669
#> [23,] 0.62626461 0.5866449 0.6405891 0.6341063 0.51835099 0.44262136 0.34470023
#> [24,] 0.24204931 0.5466100 0.6416290 0.3674985 0.46154177 0.64637119 0.65657321
#> [25,] 0.31628623 0.3522630 0.3285339 0.3137770 0.32338358 0.43382621 0.46626153
#>               V39           V4        V40        V41        V42          V43
#>  [1,]  0.36959435  0.085843482 0.34923327 0.41062884 0.46610402  0.423481007
#>  [2,]  0.67379684  0.105231067 0.46444325 0.45549242 0.32356208  0.167400584
#>  [3,]  0.36744397  0.211457940 0.31534195 0.35644941 0.48389392  0.364149416
#>  [4,]  0.37648309  0.110017058 0.24757739 0.15251201 0.14363097  0.182851700
#>  [5,]  0.19280276  0.100156546 0.14056170 0.20224461 0.32050154  0.246991428
#>  [6,]  0.43139733  0.072019075 0.39273531 0.42416616 0.41391506  0.292952378
#>  [7,]  0.22909903  0.053829183 0.23107086 0.22731408 0.26552704  0.299669764
#>  [8,]  0.17940000  0.122900000 0.02270000 0.13130000 0.17750000  0.154900000
#>  [9,]  0.24960384  0.057940508 0.26297722 0.30526971 0.27111840  0.147679033
#> [10,]  0.06370488  0.043294431 0.07555952 0.03723524 0.09943974  0.208434121
#> [11,]  0.74323186  0.106316030 0.50555687 0.28326066 0.29347913  0.267741504
#> [12,]  0.35008547  0.094300599 0.28629123 0.21729945 0.30133559  0.318632667
#> [13,]  0.15314618  0.130739592 0.13773387 0.03061385 0.07392252  0.113196891
#> [14,]  0.68407538 -0.009021324 0.52698192 0.18687524 0.19457661  0.155813062
#> [15,]  0.38058966  0.011701171 0.27788529 0.12576677 0.06699320  0.071180314
#> [16,]  0.47618467  0.008652547 0.50298756 0.45099202 0.42511989  0.272818283
#> [17,] -0.00236018  0.038374732 0.39011021 0.44684723 0.59717148  0.511836039
#> [18,]  0.12450082  0.033239960 0.13164707 0.10436428 0.04616218  0.065183947
#> [19,]  0.80523889  0.010323923 0.79585275 0.71305712 0.63918996  0.471623948
#> [20,]  0.57748268  0.132867278 0.39899868 0.21925303 0.02603564  0.195272969
#> [21,]  0.30844049  0.068908260 0.33353586 0.28749004 0.26203535  0.249505786
#> [22,]  0.11933869  0.005441872 0.13922082 0.13949450 0.13820557  0.211149968
#> [23,]  0.18695687  0.029971245 0.17039564 0.15031114 0.05671236  0.066712750
#> [24,]  0.55444064  0.074330090 0.31832777 0.22197542 0.05628270 -0.007890375
#> [25,]  0.42769633  0.012474221 0.32556825 0.19499378 0.12463985  0.082188654
#>                V44         V45          V46           V47          V48
#>  [1,]  0.368681835  0.42943442  0.345026948  0.2374155496  0.180078432
#>  [2,]  0.164704884  0.30568314  0.212536008  0.1605294028  0.225255614
#>  [3,]  0.168854739  0.17448357  0.138266770  0.0857327009  0.082840234
#>  [4,]  0.118222275  0.05283354  0.016470929 -0.0006569958  0.032793967
#>  [5,]  0.039137449  0.17593446  0.133244239  0.0463272983  0.056488239
#>  [6,]  0.284485342  0.23211572  0.033313762  0.0710815494  0.033550000
#>  [7,]  0.243520052  0.13052131  0.112070965  0.1005851572  0.113651192
#>  [8,]  0.162600000  0.07080000  0.012900000  0.0795000000  0.076200000
#>  [9,]  0.152365010  0.14057709  0.103301333  0.0905158075  0.067369868
#> [10,]  0.241086586  0.23151662  0.234298946  0.1798428767  0.109922752
#> [11,]  0.328863186  0.46264143  0.342961172  0.1318729658  0.077390058
#> [12,]  0.202499897  0.12125038  0.084641016  0.0933864574  0.055165327
#> [13,]  0.075805302  0.07323325  0.041076506  0.0415698956  0.059765052
#> [14,]  0.031821981 -0.04384097 -0.061356960  0.0409328633  0.060197802
#> [15,]  0.066036748  0.05054719  0.041015526  0.0207997393  0.017719619
#> [16,]  0.146704779  0.10184735  0.075594072  0.0926122014  0.067855457
#> [17,]  0.303345847  0.12715058  0.058398414  0.0161751065  0.056250455
#> [18,]  0.109181310  0.07637646  0.045488498  0.0633789501  0.047440238
#> [19,]  0.268747035  0.19780257  0.157526967  0.1511790594  0.103728476
#> [20,]  0.296760018  0.13000055  0.123249162  0.0395427730 -0.023476872
#> [21,]  0.229254403  0.20477374  0.120785444  0.1186149984  0.107094813
#> [22,]  0.138032303  0.01480577  0.003937409 -0.0174519843 -0.017474975
#> [23,]  0.014208912 -0.01490523  0.005049102  0.0159893541  0.027160814
#> [24,] -0.009304523 -0.01300250 -0.011815387 -0.0098778489 -0.007231975
#> [25,]  0.051613957  0.07360095  0.065709195  0.0694891950  0.051422412
#>                V49          V5           V50           V51          V52
#>  [1,]  0.107255398 0.100033351  0.0335949239  0.0298093700  0.024364871
#>  [2,]  0.111076532 0.045719751  0.0276471297  0.0410606929  0.019322524
#>  [3,]  0.046363661 0.270667318  0.0159195351  0.0260831799  0.021891931
#>  [4,]  0.012648164 0.154818771  0.0055690232  0.0054409816  0.013102226
#>  [5,]  0.048436155 0.211768446  0.0458186418  0.0224897729  0.031161614
#>  [6,]  0.005765178 0.030097831  0.0064219721  0.0189339017  0.002844669
#>  [7,]  0.082434249 0.106912944  0.0203742362  0.0280655624  0.023429927
#>  [8,]  0.011700000 0.143700000  0.0061000000  0.0257000000  0.008900000
#>  [9,]  0.033129998 0.112934512  0.0168468148  0.0149187456  0.016442097
#> [10,]  0.056428882 0.079502470  0.0179707238  0.0255006334  0.015523761
#> [11,]  0.041429597 0.094537045  0.0253130558  0.0171391859  0.024544590
#> [12,]  0.052294431 0.116182066  0.0145308541  0.0140462174  0.011554938
#> [13,]  0.038342343 0.034344110  0.0198170155  0.0327073949  0.026920059
#> [14,]  0.008111463 0.029705824  0.0098773038  0.0050121385  0.022210340
#> [15,] -0.001470744 0.033078878  0.0145748678  0.0103193628  0.006630358
#> [16,]  0.036087166 0.006785566  0.0132587486  0.0101782766  0.008146175
#> [17,]  0.002697791 0.047756037 -0.0029179976 -0.0071577671 -0.001645210
#> [18,]  0.036310066 0.053517091  0.0196771706  0.0154405896  0.009177594
#> [19,]  0.076803496 0.063685622  0.0276699287  0.0098490371  0.009435660
#> [20,]  0.025138537 0.163238410  0.0250856781  0.0170314965  0.024329161
#> [21,]  0.066429121 0.096075251  0.0254097239  0.0126482306  0.017039369
#> [22,]  0.003751237 0.025463601  0.0106151544  0.0131243678  0.011769533
#> [23,]  0.005078421 0.056051977  0.0073618917  0.0046010039  0.002981377
#> [24,] -0.003493400 0.045033133 -0.0009487319 -0.0006254992  0.004702279
#> [25,]  0.019853638 0.026522397  0.0245796892  0.0180170172  0.016437207
#>               V53         V54          V55         V56          V57
#>  [1,] 0.015149959 0.013768233 0.0115688630 0.009279174  0.009566840
#>  [2,] 0.015119362 0.033959029 0.0166104300 0.029712843  0.018200555
#>  [3,] 0.019398470 0.021594044 0.0063318836 0.013272770  0.015382035
#>  [4,] 0.006163235 0.009960971 0.0191360533 0.021456595  0.008801474
#>  [5,] 0.022434469 0.015244056 0.0225609039 0.007207416  0.004509102
#>  [6,] 0.015596811 0.012345424 0.0040649732 0.017773633  0.014428036
#>  [7,] 0.016544225 0.021800029 0.0107432318 0.006544675  0.009086806
#>  [8,] 0.026200000 0.010800000 0.0138000000 0.018700000  0.023000000
#>  [9,] 0.015929723 0.017583817 0.0073336983 0.012102657  0.008843883
#> [10,] 0.007216267 0.005238778 0.0029269743 0.003809974  0.003018180
#> [11,] 0.018208870 0.022309129 0.0232607587 0.009622284  0.009837725
#> [12,] 0.005930076 0.021822940 0.0122554116 0.008113429  0.017661172
#> [13,] 0.025960667 0.009996499 0.0177488871 0.020617016  0.002883410
#> [14,] 0.030828860 0.017211628 0.0269102910 0.001902858 -0.001565849
#> [15,] 0.006328691 0.010798760 0.0101502070 0.010703796  0.009761398
#> [16,] 0.020206379 0.013313956 0.0130145296 0.006439516  0.007867076
#> [17,] 0.010243111 0.009599821 0.0078816335 0.014874427  0.009463735
#> [18,] 0.007778699 0.006064377 0.0055634621 0.006976411  0.005059596
#> [19,] 0.017515815 0.013646273 0.0110294887 0.008888126  0.008217787
#> [20,] 0.019113748 0.010363093 0.0029263671 0.007314985  0.005199845
#> [21,] 0.009085939 0.014885929 0.0089033979 0.005239754  0.009221491
#> [22,] 0.010870704 0.003194588 0.0001830687 0.001198813  0.002331214
#> [23,] 0.008689426 0.006651841 0.0044526549 0.004600536  0.005275132
#> [24,] 0.024092746 0.020815936 0.0047092157 0.013961571  0.016613394
#> [25,] 0.022303724 0.023441132 0.0113442411 0.010133194  0.005251299
#>                 V58         V59         V6         V60         V7         V8
#>  [1,]  0.0106754861 0.009137983 0.11881431 0.007708089 0.11790026 0.13713761
#>  [2,]  0.0176357560 0.023613592 0.12219348 0.012538169 0.21021947 0.44791320
#>  [3,]  0.0132814163 0.013718944 0.17031876 0.007118710 0.07301738 0.19122289
#>  [4,]  0.0052177068 0.011941140 0.09673408 0.006492972 0.12756159 0.14677062
#>  [5,]  0.0126826727 0.015797136 0.23615158 0.009473591 0.18595826 0.14294832
#>  [6,]  0.0067401180 0.023705496 0.13647533 0.020096789 0.20083325 0.15295780
#>  [7,]  0.0206532148 0.012328769 0.18482010 0.001496782 0.18957325 0.19997209
#>  [8,]  0.0057000000 0.011300000 0.11900000 0.013100000 0.08840000 0.09070000
#>  [9,]  0.0088431245 0.006184019 0.14583159 0.005219746 0.16220777 0.18193793
#> [10,]  0.0044659102 0.002556005 0.08398271 0.003678915 0.09738332 0.12366318
#> [11,]  0.0124680751 0.008503499 0.10353708 0.014118807 0.25992664 0.41545681
#> [12,]  0.0282307078 0.013200476 0.13046444 0.004735684 0.13645553 0.10477089
#> [13,]  0.0421322637 0.013549771 0.12040609 0.016790379 0.20425972 0.15783802
#> [14,]  0.0036612995 0.008241113 0.04365081 0.008957786 0.02222945 0.05781031
#> [15,]  0.0053476025 0.006477117 0.06983492 0.006326238 0.09399684 0.07268488
#> [16,]  0.0151245905 0.020370109 0.03304953 0.001667727 0.03877115 0.04221006
#> [17,] -0.0002395742 0.003724000 0.07036692 0.016306991 0.08874794 0.12288854
#> [18,]  0.0041432021 0.004704597 0.09627828 0.003475316 0.14029130 0.09808648
#> [19,]  0.0111182722 0.008141428 0.11344988 0.004890347 0.11741557 0.10880772
#> [20,]  0.0081412547 0.001850902 0.28432646 0.005603189 0.31606635 0.23449544
#> [21,]  0.0079436981 0.010384967 0.16009151 0.007991659 0.18998188 0.20669023
#> [22,]  0.0010035404 0.003865243 0.06905733 0.006520745 0.05796917 0.04657999
#> [23,]  0.0047642962 0.002352186 0.08727176 0.002442494 0.08106609 0.09423688
#> [24,]  0.0080587094 0.013912020 0.10471962 0.010968753 0.01022327 0.06791177
#> [25,]  0.0118548078 0.002650009 0.01735341 0.003363573 0.02319770 0.03036748
#>                 V9
#>  [1,]  0.217070957
#>  [2,]  0.637350234
#>  [3,]  0.074656333
#>  [4,]  0.123230610
#>  [5,]  0.169034968
#>  [6,]  0.078033513
#>  [7,]  0.231860778
#>  [8,]  0.210700000
#>  [9,]  0.247853120
#> [10,]  0.206029462
#> [11,]  0.575098001
#> [12,]  0.170743635
#> [13,]  0.315389821
#> [14,]  0.142885471
#> [15,]  0.039687301
#> [16,] -0.019288670
#> [17,] -0.009996627
#> [18,]  0.098638895
#> [19,] -0.006151869
#> [20,]  0.323717461
#> [21,]  0.266610364
#> [22,]  0.095102496
#> [23,]  0.152359361
#> [24,]  0.128200655
#> [25,]  0.075824997
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478