Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1         V10        V11        V12        V13       V14
#>  [1,]  0.055700503 -0.08195291 0.07964709 0.27887622 0.40159801 0.4564578
#>  [2,]  0.039946700  0.16349620 0.20684961 0.20946219 0.27894367 0.3564411
#>  [3,]  0.105666715  0.34933150 0.46048915 0.54044660 0.65873851 0.7308520
#>  [4,]  0.057379112  0.53495559 0.38891046 0.12765587 0.21479148 0.3222232
#>  [5,]  0.021808774  0.23491355 0.27886429 0.35839893 0.36435594 0.3958967
#>  [6,]  0.014185648  0.30491834 0.30890247 0.32036009 0.24831401 0.1875312
#>  [7,]  0.020675260  0.35814078 0.42409909 0.43053742 0.38312554 0.2809170
#>  [8,]  0.021297681  0.30507874 0.24376967 0.30194532 0.27271815 0.2036777
#>  [9,]  0.033788969  0.26886664 0.33771870 0.32231922 0.31795683 0.2743873
#> [10,]  0.017897143  0.03764540 0.09780429 0.17033985 0.25003050 0.2704172
#> [11,]  0.051939428  0.23843051 0.33132961 0.40129901 0.38221497 0.3462313
#> [12,]  0.038546968  0.22551751 0.29639529 0.35560653 0.42029846 0.5677612
#> [13,]  0.018432765  0.19331419 0.26120934 0.36250051 0.42753215 0.5210082
#> [14,]  0.045621717  0.32944089 0.46796022 0.49728704 0.54120459 0.5680737
#> [15,]  0.028218378  0.09025713 0.14968384 0.06009541 0.09489126 0.1156046
#> [16,]  0.018097545  0.19789794 0.16401791 0.16639806 0.14490995 0.1638282
#> [17,] -0.004020352  0.11710212 0.11435028 0.14064869 0.17655122 0.1349663
#> [18,]  0.005340241  0.03052118 0.03434896 0.01700757 0.08451384 0.1171599
#> [19,]  0.020100000  0.10970000 0.08410000 0.09420000 0.12040000 0.0420000
#> [20,]  0.013766944  0.03857935 0.15201626 0.36969403 0.28209127 0.2401545
#> [21,]  0.022533963  0.17134566 0.19197158 0.22026267 0.25039505 0.3305007
#> [22,]  0.017874305  0.11121172 0.14157511 0.13015498 0.14360903 0.1281736
#> [23,]  0.016501238  0.13673402 0.11948843 0.08736075 0.06855970 0.1339068
#> [24,]  0.018619194  0.06667790 0.06976153 0.15163069 0.08840994 0.1058830
#> [25,]  0.007708404  0.10769897 0.09517760 0.09217556 0.10517003 0.1072120
#>              V15        V16       V17       V18        V19           V2
#>  [1,] 0.53647894 0.53635662 0.6753657 0.7929785 0.90873316  0.064964192
#>  [2,] 0.38288649 0.41040364 0.5496493 0.6645805 0.72919685  0.053163851
#>  [3,] 0.87677834 0.97455692 0.8860307 0.8788104 0.90839103  0.112239452
#>  [4,] 0.42767753 0.42385596 0.1669616 0.1424515 0.32274745  0.087954403
#>  [5,] 0.42182437 0.49807277 0.5918111 0.6810382 0.73189516  0.019795271
#>  [6,] 0.12362091 0.23059341 0.2976516 0.3430655 0.24581053  0.038939304
#>  [7,] 0.20210076 0.20850517 0.2559331 0.3224419 0.51289253  0.025486994
#>  [8,] 0.11012848 0.14711112 0.1612302 0.2005845 0.16446338  0.023463523
#>  [9,] 0.20277575 0.18204180 0.2456903 0.2642417 0.36654019  0.050014636
#> [10,] 0.30668612 0.46911159 0.5549966 0.6292684 0.84125109 -0.000647270
#> [11,] 0.30055088 0.32340838 0.3042157 0.2700750 0.28392649  0.061612524
#> [12,] 0.78744383 0.81807736 0.7851115 0.8923839 1.00517322  0.054117420
#> [13,] 0.55993534 0.53497334 0.6598202 0.7058675 0.79788763  0.028369727
#> [14,] 0.67338617 0.80014240 0.9262125 0.9489034 0.87327494  0.068120660
#> [15,] 0.15509220 0.17523345 0.1562626 0.1299493 0.07390518  0.021932012
#> [16,] 0.22314626 0.28055252 0.3975566 0.5217867 0.47317665  0.016442248
#> [17,] 0.20801456 0.30084259 0.3750109 0.3569755 0.18815539 -0.006473670
#> [18,] 0.15349550 0.15818103 0.2100472 0.2657375 0.38320378  0.001611762
#> [19,] 0.00310000 0.01620000 0.0624000 0.2127000 0.34360000  0.002600000
#> [20,] 0.10567936 0.08869154 0.1843923 0.4337290 0.28034670  0.045469164
#> [21,] 0.36305256 0.51657213 0.6345133 0.5392539 0.39605755  0.032503202
#> [22,] 0.06928632 0.09441316 0.1284266 0.1837048 0.36123880  0.013637868
#> [23,] 0.07371515 0.09455546 0.2441752 0.2805692 0.32913114  0.007550487
#> [24,] 0.09318518 0.16749223 0.1312949 0.1689545 0.24943000  0.027239344
#> [25,] 0.15603366 0.17040697 0.1336295 0.1455104 0.11733684  0.012339000
#>             V20        V21       V22       V23       V24        V25       V26
#>  [1,] 0.9432766 0.87440836 0.8712517 0.9461407 1.0207816 0.83776505 0.5513306
#>  [2,] 0.7683437 0.87622094 0.8967710 0.9216418 0.9532762 0.96770347 1.0351015
#>  [3,] 0.9501555 0.85906533 0.7910633 0.8458823 0.7738585 0.56005773 0.5186900
#>  [4,] 0.8951775 0.94994555 0.3835927 0.1560407 0.1851705 0.14156610 0.3618881
#>  [5,] 0.8231557 0.86553310 0.9144443 0.9635304 0.9837799 0.98895205 0.9659185
#>  [6,] 0.1495734 0.07852028 0.1218770 0.2935424 0.4214038 0.57857098 0.7805583
#>  [7,] 0.6443072 0.78484635 0.8405648 0.8303387 0.8966306 0.90829058 0.9331687
#>  [8,] 0.1269410 0.03977704 0.2064339 0.4502724 0.6067280 0.71713721 0.8061670
#>  [9,] 0.4401222 0.56825961 0.6816755 0.6886133 0.7119823 0.69736841 0.6962285
#> [10,] 0.9750516 0.95368177 0.9808108 1.0378871 0.9473674 0.89058583 0.8354877
#> [11,] 0.5333423 0.61436212 0.5854574 0.5871994 0.5925856 0.57564268 0.7429871
#> [12,] 0.9287521 0.90723650 0.9525254 0.9314510 0.7819213 0.62427227 0.5732625
#> [13,] 0.9278411 1.04756103 1.0077771 0.8072383 0.6834389 0.67785111 0.7811239
#> [14,] 0.8033489 0.73299563 0.5134445 0.2603870 0.1689421 0.27031438 0.2578601
#> [15,] 0.1658861 0.32382683 0.5791966 0.8618519 0.8853528 0.84640758 0.8653149
#> [16,] 0.5941570 0.66363406 0.6536278 0.6890183 0.6310299 0.42446312 0.4441595
#> [17,] 0.1604985 0.35005026 0.2881594 0.2905163 0.5291870 0.84554470 0.9446394
#> [18,] 0.4666389 0.50520165 0.6217281 0.6942559 0.8005146 0.90858993 0.8607098
#> [19,] 0.3813000 0.38250000 0.4764000 0.6313000 0.7523000 0.86750000 0.8788000
#> [20,] 0.5869861 0.74335177 0.7468170 0.7623936 0.8293270 1.01664398 0.9303793
#> [21,] 0.3483902 0.40749106 0.4874440 0.6016481 0.7288821 0.78998335 0.7488574
#> [22,] 0.3419900 0.23134824 0.2762439 0.4218308 0.5813172 0.57459563 0.5653740
#> [23,] 0.4819306 0.61172147 0.7495837 0.8850883 0.8951825 0.89970656 0.9707704
#> [24,] 0.2528591 0.25674383 0.1508919 0.2326422 0.2393426 0.09446879 0.2632014
#> [25,] 0.1339964 0.21901210 0.2751870 0.4118341 0.6147477 0.83085069 0.9714705
#>             V27       V28       V29           V3       V30       V31       V32
#>  [1,] 0.3935561 0.3444806 0.3560659  0.088358520 0.2778658 0.2484342 0.3028370
#>  [2,] 1.0040512 1.0366084 0.9272063  0.031199816 0.7243784 0.5535999 0.3079996
#>  [3,] 0.5384020 0.5931160 0.5279894  0.084456514 0.4214201 0.3447777 0.2665384
#>  [4,] 0.2726507 0.5341994 0.4854186  0.094977014 0.7314932 0.2953702 0.3496325
#>  [5,] 0.9470468 0.9122621 0.8115431  0.026688983 0.6175887 0.3778074 0.3331689
#>  [6,] 0.9125428 0.9240852 0.9309170  0.025864862 1.0011785 0.9413777 0.8386736
#>  [7,] 0.9693387 0.9446939 0.7838276  0.033277990 0.5659005 0.4147218 0.3787705
#>  [8,] 0.9285264 0.9583448 0.8863539  0.028400765 0.9125026 0.9388894 0.9012453
#>  [9,] 0.7533229 0.7926745 0.7925876  0.059485421 0.7394269 0.7311094 0.6170317
#> [10,] 0.8105066 0.7883237 0.6400386  0.009693478 0.4375787 0.2237284 0.1406838
#> [11,] 0.8813884 0.8777736 0.7388067  0.070977760 0.6302104 0.5873094 0.6587477
#> [12,] 0.4366867 0.3338523 0.5240129  0.052711085 0.3989085 0.3134107 0.1443725
#> [13,] 0.7854344 0.6590558 0.5127579  0.027222321 0.4407420 0.2373745 0.1634438
#> [14,] 0.2303106 0.2035596 0.0532956  0.095872604 0.2195529 0.2081794 0.2526230
#> [15,] 0.8237183 0.8565690 0.9467900  0.025656100 0.7895140 0.6232477 0.5043124
#> [16,] 0.6171545 0.7935983 0.8031985  0.014123930 0.7328792 0.7406717 0.6991261
#> [17,] 0.7727574 0.7146083 0.5963289 -0.011106676 0.3739280 0.2969336 0.2164412
#> [18,] 0.8982903 0.8135816 0.5787782 -0.003288011 0.5145735 0.4369029 0.3626228
#> [19,] 0.7901000 0.8357000 0.9631000  0.013800000 0.9619000 0.9236000 0.8903000
#> [20,] 0.7053219 0.5184301 0.4839441  0.063316863 0.4478449 0.5302259 0.5382467
#> [21,] 0.7370666 0.7403251 0.7050796  0.037577749 0.7118463 0.6961305 0.6070376
#> [22,] 0.5071664 0.7051572 0.9755520  0.017505242 0.9932872 0.9339571 0.8626436
#> [23,] 0.9980780 0.9173671 0.8493456  0.013149395 0.7307443 0.6435939 0.5991367
#> [24,] 0.4532326 0.6076434 0.6826212  0.038500793 0.6837831 0.6822296 0.4927222
#> [25,] 0.8285382 0.5961241 0.4854611  0.007898444 0.3611162 0.2801456 0.1458302
#>              V33        V34        V35       V36        V37        V38
#>  [1,] 0.48394303 0.51673587 0.45445842 0.3335316 0.32509964 0.47201091
#>  [2,] 0.23338136 0.18335645 0.21965068 0.1273541 0.04626320 0.16567061
#>  [3,] 0.15840300 0.19298167 0.18061025 0.1440778 0.18484592 0.17720114
#>  [4,] 0.60423118 0.78863428 0.87575295 0.9039471 0.83695925 0.93136135
#>  [5,] 0.20677011 0.07719919 0.10883435 0.1410746 0.21152798 0.30363518
#>  [6,] 0.73411247 0.67758481 0.73403848 0.7918247 0.77830233 0.63116150
#>  [7,] 0.34394928 0.27602941 0.17359299 0.1837708 0.20353150 0.24529004
#>  [8,] 0.76749609 0.66016889 0.64633754 0.6805017 0.68195896 0.55827649
#>  [9,] 0.54993114 0.51429129 0.42496766 0.3243003 0.25594160 0.21850423
#> [10,] 0.17284629 0.26765736 0.37540436 0.3838899 0.30538602 0.21882950
#> [11,] 0.50689512 0.36626181 0.25230446 0.3255408 0.31580952 0.43406960
#> [12,] 0.08691543 0.16522220 0.21146791 0.1741096 0.14653087 0.07298629
#> [13,] 0.13264854 0.04127336 0.04207174 0.1886325 0.22073038 0.33340973
#> [14,] 0.29875926 0.27427262 0.34575736 0.1205831 0.14971834 0.28734693
#> [15,] 0.56404056 0.72541906 0.71920527 0.5472935 0.28961491 0.17342920
#> [16,] 0.52674968 0.41319633 0.42374078 0.5240019 0.50109581 0.35545554
#> [17,] 0.19336222 0.21290524 0.32705791 0.2292077 0.07303060 0.06001143
#> [18,] 0.27421126 0.19066443 0.18082994 0.2774924 0.32714798 0.33273984
#> [19,] 0.97080000 0.96470000 0.78920000 0.5307000 0.27180000 0.19530000
#> [20,] 0.65038757 0.54821442 0.55117585 0.5596961 0.49874092 0.21907313
#> [21,] 0.54548149 0.59071991 0.76440327 0.7788494 0.67749181 0.48774653
#> [22,] 0.76328314 0.64953333 0.54524735 0.4620075 0.46217153 0.22314701
#> [23,] 0.55960366 0.59072132 0.61449427 0.5548856 0.49149281 0.39862922
#> [24,] 0.57795182 0.70238721 0.70840615 0.7098272 0.67189832 0.67723557
#> [25,] 0.18137746 0.21193943 0.24260456 0.1807571 0.06021104 0.06871068
#>               V39           V4        V40       V41        V42       V43
#>  [1,]  0.44647604  0.058383526 0.42143508 0.3824135 0.40994987 0.3197538
#>  [2,]  0.18169708  0.015830176 0.24943675 0.1153182 0.09080569 0.2062633
#>  [3,]  0.20986547  0.037521391 0.25982206 0.1818453 0.16029916 0.1608350
#>  [4,]  0.73585279  0.100344608 0.48684826 0.4123151 0.31540931 0.1995639
#>  [5,]  0.28797290  0.045728816 0.28678635 0.3222789 0.29417776 0.2120189
#>  [6,]  0.50563623  0.066711891 0.49919509 0.4305638 0.37367726 0.4492023
#>  [7,]  0.31655550  0.036382310 0.27977793 0.1207033 0.16211251 0.2636225
#>  [8,]  0.49282187  0.034772108 0.54517260 0.5264083 0.43575747 0.4580932
#>  [9,]  0.27332593  0.062578785 0.36379009 0.4529388 0.46259905 0.3233813
#> [10,]  0.35204641  0.101814292 0.42250509 0.4364756 0.43319407 0.3645171
#> [11,]  0.52590243  0.040650330 0.38160218 0.4670004 0.65993657 0.5873636
#> [12,]  0.28582526  0.042646476 0.27814518 0.1346689 0.14413034 0.2983173
#> [13,]  0.41888311  0.033878059 0.25700967 0.1429366 0.22301486 0.1774268
#> [14,]  0.22258033  0.106111388 0.08473189 0.1406692 0.22008923 0.1659286
#> [15,]  0.17926920  0.048603963 0.26359207 0.2151068 0.22422833 0.3089534
#> [16,]  0.33099596  0.010869377 0.29421061 0.2521024 0.17622752 0.1369788
#> [17,]  0.05604792  0.001021867 0.10226097 0.1818596 0.17007068 0.1734620
#> [18,]  0.28182659 -0.016576688 0.19821958 0.1608130 0.15468189 0.1604581
#> [19,]  0.13740000  0.006200000 0.31050000 0.3790000 0.41050000 0.3355000
#> [20,] -0.01420073  0.050264181 0.33582430 0.4204099 0.58939577 0.4964719
#> [21,]  0.37965151  0.038951069 0.36678321 0.2970578 0.23351255 0.1915986
#> [22,]  0.10573302  0.039053261 0.16867430 0.1749306 0.15498492 0.2413663
#> [23,]  0.22051995  0.023236229 0.18182507 0.1799930 0.07564087 0.0632992
#> [24,]  0.69192714  0.022117021 0.64986679 0.4849154 0.41301120 0.3018392
#> [25,]  0.17397894  0.015644284 0.23842290 0.2948779 0.29506808 0.2192691
#>              V44        V45        V46           V47         V48          V49
#>  [1,] 0.27027598 0.21001667 0.14638660  0.0716100065 0.015076040  0.032116978
#>  [2,] 0.17763483 0.20399798 0.11549226  0.0753210887 0.049935126  0.053854049
#>  [3,] 0.21430885 0.12846171 0.07220943  0.0857825977 0.085900574  0.067728292
#>  [4,] 0.23520567 0.38357632 0.28327760  0.1684845267 0.191915384  0.089094984
#>  [5,] 0.19700006 0.13898251 0.11890257  0.1219621485 0.089814461  0.054314781
#>  [6,] 0.52063669 0.44186696 0.25736914  0.1411385631 0.143402064  0.093755880
#>  [7,] 0.30412504 0.28446949 0.20107062  0.1396344620 0.101964583  0.056106585
#>  [8,] 0.53386481 0.49144706 0.38834512  0.2996202087 0.219645228  0.111580943
#>  [9,] 0.20197860 0.26485021 0.22617144  0.1938232631 0.155702192  0.076735109
#> [10,] 0.25770592 0.18894006 0.12705292  0.1313466195 0.054777747  0.027418249
#> [11,] 0.47783558 0.53980526 0.40522509  0.2784750077 0.218634205  0.127104992
#> [12,] 0.20994026 0.10889747 0.13487153  0.1014053399 0.046005202  0.037701457
#> [13,] 0.15922043 0.11581770 0.08973660  0.0976828640 0.093569816  0.071088703
#> [14,] 0.09791233 0.09277226 0.05137262  0.0604110309 0.056070976  0.020104058
#> [15,] 0.18208774 0.06134271 0.05883279  0.0639786894 0.051731544  0.009365460
#> [16,] 0.11147295 0.08520665 0.04570258  0.0383990604 0.038589667  0.031370579
#> [17,] 0.11175834 0.07544011 0.06272524  0.0005144741 0.025326652  0.020065468
#> [18,] 0.15914164 0.09275836 0.02665929 -0.0137012190 0.002626799 -0.001572709
#> [19,] 0.29980000 0.27480000 0.20240000  0.1043000000 0.045300000  0.033700000
#> [20,] 0.30267287 0.15650250 0.10148039  0.0649566361 0.070756087  0.014636480
#> [21,] 0.16708932 0.16672824 0.14771997  0.0916470831 0.061026092  0.031827470
#> [22,] 0.35512374 0.34906045 0.22206622  0.0999190450 0.024872710  0.015935939
#> [23,] 0.04700986 0.03658641 0.02989089  0.0207061896 0.029442146  0.010099953
#> [24,] 0.21210689 0.17826675 0.11126340  0.1101326338 0.085801394  0.045275159
#> [25,] 0.13227726 0.10412332 0.11183600  0.0558482883 0.033794298  0.028329799
#>                 V5         V50          V51           V52         V53
#>  [1,]  0.126879387 0.006680084  0.001077840  0.0108614348 0.015534536
#>  [2,]  0.157338934 0.019269001  0.027355975  0.0118134960 0.020444617
#>  [3,]  0.086179776 0.012500181  0.022336037  0.0163245814 0.013845137
#>  [4,]  0.055292216 0.025327724  0.034041253  0.0238244107 0.015265886
#>  [5,]  0.090814567 0.026023875  0.014321978  0.0148742551 0.016373440
#>  [6,]  0.097622873 0.025223870  0.011423926  0.0179397084 0.012710486
#>  [7,]  0.066444352 0.014979393  0.021001186  0.0117985745 0.007364945
#>  [8,]  0.066339646 0.038289258  0.022175387  0.0135996781 0.005817862
#>  [9,]  0.072708446 0.017580290  0.021778515  0.0204651138 0.011511982
#> [10,]  0.171546426 0.021834073  0.012979938  0.0139618950 0.014173462
#> [11,]  0.083393510 0.039503221  0.045519992  0.0290140794 0.009589034
#> [12,]  0.004798342 0.018038157  0.037236680  0.0128213681 0.005660825
#> [13,]  0.124878168 0.028996411  0.021198517  0.0191408669 0.011231593
#> [14,]  0.118666210 0.016951596  0.024908590  0.0169091836 0.023916544
#> [15,]  0.067438034 0.008626864  0.009143588  0.0040898595 0.004879875
#> [16,]  0.011322424 0.011413411  0.006142653  0.0049425040 0.004072234
#> [17,]  0.008223853 0.010038217  0.005761413  0.0116454078 0.005684491
#> [18,] -0.014190959 0.007411551 -0.000149240  0.0030837253 0.011062428
#> [19,]  0.013300000 0.012200000  0.007200000  0.0108000000 0.007000000
#> [20,]  0.119157669 0.004239329  0.002097641 -0.0009776105 0.011453010
#> [21,]  0.057106104 0.019867543  0.013407078  0.0115958561 0.008266548
#> [22,]  0.038040592 0.017532790  0.011672950  0.0032060999 0.007148438
#> [23,]  0.048546914 0.007822508  0.005028118  0.0048963238 0.009133776
#> [24,]  0.032863884 0.016606087  0.009485467  0.0086582324 0.016562612
#> [25,]  0.030132642 0.009363617  0.004321842  0.0076602503 0.003314733
#>               V54         V55         V56         V57         V58         V59
#>  [1,] 0.014633175 0.007210408 0.015744379 0.011773918 0.013690241 0.014780508
#>  [2,] 0.004949917 0.009714073 0.004168535 0.006857108 0.012978907 0.007912426
#>  [3,] 0.017424015 0.017837696 0.005030716 0.003800816 0.010018680 0.010413977
#>  [4,] 0.024255006 0.016728953 0.022479299 0.016948132 0.016367957 0.018131605
#>  [5,] 0.014045060 0.007651144 0.014669251 0.009480386 0.008851029 0.006012640
#>  [6,] 0.010348141 0.002956916 0.007480173 0.006801675 0.005545743 0.001612200
#>  [7,] 0.009361173 0.006626625 0.005612891 0.004421394 0.006218388 0.006794341
#>  [8,] 0.007808030 0.007977473 0.008037641 0.007242073 0.002496446 0.003269432
#>  [9,] 0.012950497 0.008312316 0.007475052 0.005770659 0.009121624 0.009473860
#> [10,] 0.007444575 0.001366780 0.010223501 0.008569360 0.011246627 0.017403305
#> [11,] 0.012975993 0.017025716 0.007688860 0.008072732 0.007725303 0.009096211
#> [12,] 0.023695932 0.001832276 0.008545659 0.011764640 0.011327062 0.009831805
#> [13,] 0.023510320 0.012272091 0.006663277 0.007798500 0.008830989 0.008577497
#> [14,] 0.014873822 0.018283838 0.014025030 0.012439387 0.014423630 0.012258044
#> [15,] 0.006881934 0.004812044 0.007013947 0.013355070 0.011331139 0.005519969
#> [16,] 0.005468422 0.007323435 0.004887827 0.007771003 0.007134532 0.004321767
#> [17,] 0.007282567 0.008441305 0.004486638 0.004244247 0.004520559 0.002436466
#> [18,] 0.011010103 0.003384162 0.003555384 0.003094856 0.005551093 0.002555837
#> [19,] 0.006300000 0.003000000 0.001100000 0.000700000 0.002400000 0.005700000
#> [20,] 0.010010554 0.009853521 0.011496921 0.011228058 0.005791525 0.003080327
#> [21,] 0.012254470 0.010926584 0.006311788 0.011559592 0.006665673 0.009603934
#> [22,] 0.011438857 0.001860563 0.002900108 0.005747649 0.003885998 0.006554529
#> [23,] 0.005741243 0.003382983 0.004885286 0.003389488 0.003609075 0.003345728
#> [24,] 0.015334510 0.010412092 0.008255217 0.007789394 0.009357059 0.012567574
#> [25,] 0.003393025 0.005368082 0.003248397 0.002525266 0.003587409 0.001298571
#>                 V6         V60         V7         V8          V9
#>  [1,]  0.168504734 0.007935704 0.18125184 0.01365783 0.006180122
#>  [2,]  0.216563486 0.003496597 0.14436670 0.18077791 0.192689515
#>  [3,]  0.162972607 0.011605416 0.17625145 0.17220196 0.241009208
#>  [4,]  0.130628270 0.014505822 0.24833058 0.41505377 0.607372717
#>  [5,]  0.132427355 0.004603591 0.14176546 0.13143462 0.198803231
#>  [6,]  0.104551521 0.002966077 0.15058709 0.20023590 0.309837549
#>  [7,]  0.103966796 0.005796555 0.12723798 0.16929931 0.271104924
#>  [8,]  0.089517443 0.002653035 0.12099333 0.12696429 0.246885422
#>  [9,]  0.124733277 0.006514675 0.09219380 0.10132639 0.209006366
#> [10,]  0.025913812 0.013758203 0.05983290 0.06500011 0.016171789
#> [11,]  0.117771072 0.005428565 0.10095422 0.16660119 0.210207922
#> [12,]  0.138482398 0.012261445 0.19979392 0.26180835 0.294974251
#> [13,]  0.167584624 0.007349025 0.16131316 0.15517571 0.207948072
#> [14,]  0.096320377 0.011913599 0.10531703 0.07387394 0.164699428
#> [15,]  0.094194290 0.005506895 0.08696680 0.11090047 0.075564025
#> [16,]  0.081849712 0.003283640 0.08555718 0.15492904 0.173148446
#> [17,]  0.085308590 0.001191123 0.09718019 0.09681860 0.103612626
#> [18,] -0.003897389 0.004105344 0.03604059 0.04922441 0.051834315
#> [19,]  0.015100000 0.004400000 0.05410000 0.02100000 0.050500000
#> [20,]  0.179229282 0.014750972 0.13969710 0.12416426 0.009996958
#> [21,]  0.112889376 0.007324069 0.14202611 0.13326670 0.141108416
#> [22,]  0.025778206 0.001442679 0.08376160 0.08873219 0.078399953
#> [23,]  0.082645756 0.003130814 0.08227171 0.07432892 0.150822351
#> [24,]  0.059468823 0.005244346 0.04919924 0.08603550 0.057473227
#> [25,]  0.044379558 0.003108782 0.06338962 0.06298401 0.107239349
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3768116