Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1       V10        V11        V12       V13       V14
#>  [1,] 0.009843857 0.3377792 0.30704797 0.31708627 0.2454329 0.1458391
#>  [2,] 0.048366159 0.2647642 0.26457717 0.28038430 0.3188521 0.3008915
#>  [3,] 0.085904231 0.6412531 0.49396766 0.25317975 0.1504733 0.4012678
#>  [4,] 0.035992112 0.3795235 0.43439914 0.44859086 0.5010473 0.4253567
#>  [5,] 0.041094157 0.1466917 0.24034847 0.26887250 0.2237000 0.2179580
#>  [6,] 0.029564731 0.2877006 0.32173622 0.32500600 0.2942155 0.2880142
#>  [7,] 0.030635837 0.2558389 0.41425059 0.52480128 0.5760412 0.5761117
#>  [8,] 0.029949717 0.2498516 0.29699029 0.29659872 0.2301564 0.1113181
#>  [9,] 0.026594054 0.1162402 0.19046517 0.25512985 0.3075267 0.4011006
#> [10,] 0.023398484 0.3797022 0.49289495 0.50315352 0.4210484 0.3012088
#> [11,] 0.080078328 0.5213179 0.37988397 0.16162970 0.2337605 0.3200267
#> [12,] 0.021064574 0.2369792 0.29945038 0.31241811 0.2720685 0.2122957
#> [13,] 0.091661288 0.2956305 0.39787996 0.41976951 0.5966501 0.7900332
#> [14,] 0.025793965 0.2211418 0.25568330 0.30618563 0.3179710 0.3493620
#> [15,] 0.008480765 0.2090602 0.15416636 0.08887356 0.2214328 0.3030543
#> [16,] 0.015698775 0.1258714 0.15231060 0.14741144 0.1765203 0.1984533
#> [17,] 0.017512921 0.1077488 0.13277630 0.11481306 0.1339862 0.1228849
#> [18,] 0.016942232 0.1329758 0.10990939 0.06522228 0.1741682 0.3465842
#> [19,] 0.015578876 0.1348898 0.13333725 0.16327661 0.2313592 0.3128714
#> [20,] 0.029300000 0.0860000 0.04140000 0.04720000 0.0835000 0.0938000
#> [21,] 0.015442165 0.1004285 0.10700266 0.10984170 0.1259097 0.1544336
#> [22,] 0.006624584 0.1072043 0.05124833 0.15998969 0.1775241 0.1043654
#> [23,] 0.004941224 0.0488250 0.05683564 0.11122700 0.1446560 0.1497497
#> [24,] 0.020328043 0.0614067 0.05705752 0.16207871 0.1314360 0.1084925
#> [25,] 0.025711363 0.1141646 0.16164268 0.16680489 0.1433767 0.2108092
#>              V15       V16        V17        V18       V19          V2
#>  [1,] 0.17863982 0.2152885 0.17963185 0.16899803 0.1336070 0.009293779
#>  [2,] 0.32651510 0.3316751 0.31665709 0.16728200 0.3110057 0.060578077
#>  [3,] 0.45608737 0.4597535 0.29139814 0.32699587 0.4845874 0.083929248
#>  [4,] 0.27093441 0.3005120 0.32934484 0.31723794 0.5158837 0.045913221
#>  [5,] 0.20055413 0.1761245 0.20177930 0.28110129 0.3816093 0.062827455
#>  [6,] 0.21587466 0.1694067 0.24475691 0.32752679 0.5107467 0.038522677
#>  [7,] 0.62417216 0.7424810 0.89367747 0.96405953 0.9554027 0.037190939
#>  [8,] 0.12662755 0.1859729 0.22489834 0.25881273 0.2405316 0.027976631
#>  [9,] 0.54472510 0.6253578 0.73505124 0.83012929 0.9347563 0.024583201
#> [10,] 0.23809184 0.3429708 0.38346809 0.38822555 0.3986585 0.029091106
#> [11,] 0.45800405 0.5239056 0.20362324 0.15750632 0.2528835 0.082234674
#> [12,] 0.16713090 0.2006210 0.23778955 0.36739469 0.5807663 0.049116616
#> [13,] 0.92480762 0.9624926 0.84863858 0.88166320 0.9342729 0.100022123
#> [14,] 0.41949671 0.5232121 0.58363809 0.66501255 0.7196554 0.032594655
#> [15,] 0.38487730 0.2539516 0.09279012 0.15067374 0.2170983 0.019710543
#> [16,] 0.19022784 0.2134568 0.27866925 0.30630452 0.3861873 0.011746547
#> [17,] 0.09182751 0.1310764 0.17242601 0.23355666 0.3970164 0.014398708
#> [18,] 0.51317119 0.7032794 0.78480134 0.78761102 0.7915412 0.017980131
#> [19,] 0.27820486 0.3224970 0.44120258 0.67162849 0.7763574 0.022997141
#> [20,] 0.14660000 0.0809000 0.11790000 0.21790000 0.3326000 0.064400000
#> [21,] 0.19799717 0.1908959 0.16505356 0.16652521 0.1061595 0.016445627
#> [22,] 0.15292304 0.1427654 0.11936752 0.21757234 0.1076375 0.017693402
#> [23,] 0.09264841 0.1271814 0.16231836 0.16979828 0.1465181 0.008652394
#> [24,] 0.09808599 0.1648871 0.16658524 0.09996505 0.1468248 0.027992273
#> [25,] 0.29816258 0.3386917 0.41714339 0.41156310 0.3135004 0.035976421
#>             V20       V21       V22        V23        V24        V25       V26
#>  [1,] 0.1048061 0.1247656 0.2851418 0.48217206 0.63360280 0.74990397 0.8488612
#>  [2,] 0.4310408 0.4869277 0.4494917 0.46624739 0.48873109 0.55018087 0.5803848
#>  [3,] 0.9009305 0.9793733 0.3959377 0.10893830 0.05621657 0.24150763 0.4358184
#>  [4,] 0.7354343 0.7685679 0.7768614 0.70877658 0.69383677 0.70797387 0.8510329
#>  [5,] 0.3651788 0.3777418 0.4092185 0.55780868 0.64062131 0.64550563 0.7335216
#>  [6,] 0.6092474 0.7118272 0.6646604 0.55939309 0.56626426 0.65696795 0.7226063
#>  [7,] 0.9161161 0.8026145 0.6155538 0.44652374 0.38590435 0.32459930 0.1624497
#>  [8,] 0.2780403 0.4387115 0.6503501 0.77284486 0.82196075 0.91808925 0.9697872
#>  [9,] 0.9876222 1.0070634 1.0119313 0.94236002 0.80205037 0.59230331 0.5012655
#> [10,] 0.5084930 0.7054706 0.9281922 0.97508762 0.95102423 0.93390341 0.8564464
#> [11,] 0.8582747 0.9575599 0.3826276 0.13119224 0.18568723 0.20251832 0.3538157
#> [12,] 0.6915172 0.7870562 0.7926823 0.81261537 0.92294780 0.92463078 0.9443751
#> [13,] 0.9310553 0.8679451 0.8518208 0.93039943 0.86433961 0.64676558 0.6104973
#> [14,] 0.8093632 0.8477498 0.8892105 0.92222579 0.96675832 0.97977107 0.9644157
#> [15,] 0.2334509 0.2318485 0.3740219 0.60662997 0.76630499 0.70128774 0.7113129
#> [16,] 0.4847748 0.5515304 0.6348400 0.75964136 0.80467351 0.84877168 0.8776365
#> [17,] 0.3796847 0.2831524 0.2973884 0.39726826 0.53919653 0.53506510 0.5390871
#> [18,] 0.7885909 0.8593062 0.8718063 0.90563520 0.93816493 0.91634777 0.9268544
#> [19,] 0.8522597 0.8773753 0.8406818 0.82231658 0.76525240 0.65193340 0.4884304
#> [20,] 0.3258000 0.2111000 0.2302000 0.33610000 0.42590000 0.46090000 0.2606000
#> [21,] 0.1357995 0.2164764 0.2104909 0.33492523 0.58131482 0.85869009 0.9999123
#> [22,] 0.1340397 0.2698243 0.1464082 0.08013421 0.13996965 0.07019264 0.1770921
#> [23,] 0.1470549 0.2319985 0.3901842 0.44086090 0.55756901 0.65697380 0.7301499
#> [24,] 0.1948568 0.1941919 0.1157121 0.18980945 0.16163373 0.10170520 0.2600260
#> [25,] 0.3553034 0.4372130 0.5780043 0.72035314 0.81776172 0.83460642 0.7729330
#>             V27       V28       V29         V3       V30       V31       V32
#>  [1,] 0.9191949 0.9360923 0.8912314 0.01377770 0.9200867 0.9130281 0.8453608
#>  [2,] 0.6378824 0.7977680 0.8443917 0.07428598 0.8757060 0.8841247 0.7342011
#>  [3,] 0.2582201 0.5453664 0.6567153 0.04849875 0.8352606 0.4629972 0.4135329
#>  [4,] 0.9510298 0.9171947 0.7817316 0.05007051 0.6084239 0.5254526 0.6071610
#>  [5,] 0.8266752 0.7787604 0.7273577 0.09201179 0.6184103 0.4919563 0.3833127
#>  [6,] 0.8407310 0.8106762 0.8880481 0.04500269 0.9178720 0.7537952 0.5772382
#>  [7,] 0.1481608 0.1546215 0.1136111 0.05826795 0.1483164 0.2402458 0.3240625
#>  [8,] 0.9870090 0.9601806 0.8302898 0.02755503 0.7051893 0.6716417 0.5479611
#>  [9,] 0.4928151 0.5148899 0.5469642 0.03623631 0.4432651 0.2870470 0.2614194
#> [10,] 0.7538826 0.6057155 0.4081569 0.02779696 0.3620030 0.3427789 0.3947238
#> [11,] 0.2778030 0.4485007 0.4327787 0.06687343 0.5992287 0.2524107 0.3398856
#> [12,] 0.9807848 0.9708814 0.8511947 0.03951954 0.6245758 0.4034408 0.3033219
#> [13,] 0.5485600 0.5217638 0.5600439 0.06791665 0.3885289 0.3037359 0.2071971
#> [14,] 0.9195017 0.8415214 0.7394512 0.02549515 0.5899983 0.4297527 0.3283079
#> [15,] 0.7286507 0.8677712 0.9704823 0.04521267 0.9163203 0.8439625 0.7085201
#> [16,] 0.9126966 0.8816463 0.7461119 0.02044141 0.6263530 0.5388229 0.4366864
#> [17,] 0.5462510 0.7544291 0.9847422 0.01698227 0.9772479 0.9232290 0.8521041
#> [18,] 0.9044868 0.7341776 0.4879550 0.02767433 0.4185286 0.4153603 0.3863038
#> [19,] 0.3230243 0.2272740 0.3794971 0.01025632 0.3123748 0.3373164 0.2950079
#> [20,] 0.0874000 0.2862000 0.5606000 0.03900000 0.8344000 0.8096000 0.7250000
#> [21,] 0.8482697 0.6472500 0.5373802 0.01731643 0.4133391 0.3222738 0.1733097
#> [22,] 0.2864558 0.4946300 0.5632877 0.01572833 0.5720351 0.5875573 0.6028132
#> [23,] 0.7456939 0.8317061 0.9491316 0.01334894 0.9371206 0.7684407 0.7044140
#> [24,] 0.5067332 0.6937973 0.7119213 0.03217477 0.6814052 0.7228666 0.5478492
#> [25,] 0.6949271 0.6950046 0.6696705 0.03616682 0.6185547 0.6226790 0.5920955
#>              V33       V34        V35        V36        V37        V38
#>  [1,] 0.72768882 0.6574517 0.63799858 0.67665879 0.63223699 0.51688412
#>  [2,] 0.48069278 0.4013670 0.34654845 0.18557099 0.19898928 0.28738283
#>  [3,] 0.56007706 0.7628283 0.80867201 0.74594367 0.73720192 0.83769054
#>  [4,] 0.52794253 0.3335235 0.10676266 0.19026999 0.19940055 0.36734047
#>  [5,] 0.33681921 0.2717662 0.15223471 0.16548030 0.27201540 0.24937100
#>  [6,] 0.60085462 0.5518900 0.38799917 0.33588977 0.29373957 0.19155861
#>  [7,] 0.32832596 0.3914769 0.41575688 0.23007745 0.22692941 0.31176549
#>  [8,] 0.40429834 0.1789546 0.09533035 0.07091243 0.11797112 0.15782013
#>  [9,] 0.34656299 0.3219428 0.29434899 0.27708155 0.26448024 0.31044016
#> [10,] 0.44848486 0.3999984 0.22478934 0.28719752 0.35005459 0.30616420
#> [11,] 0.56547246 0.7155327 0.77369479 0.81526782 0.79465320 0.84400529
#> [12,] 0.25047512 0.2066433 0.17772644 0.07981340 0.06839823 0.09645422
#> [13,] 0.09901746 0.1657369 0.18965275 0.15379446 0.14966032 0.12036541
#> [14,] 0.25500254 0.1544354 0.13883982 0.13533808 0.22848423 0.29680767
#> [15,] 0.65061339 0.5629322 0.45251374 0.43815623 0.26280918 0.13816919
#> [16,] 0.32495428 0.2574627 0.22670473 0.29496989 0.27861474 0.26904801
#> [17,] 0.75626460 0.6448888 0.55048737 0.48821200 0.48392083 0.24055286
#> [18,] 0.27642795 0.2880908 0.29351580 0.21522201 0.22570769 0.20441450
#> [19,] 0.30689399 0.2916456 0.30838517 0.46197742 0.42924127 0.33191789
#> [20,] 0.80480000 0.9435000 1.00000000 0.89600000 0.55160000 0.30370000
#> [21,] 0.16337647 0.2009970 0.27759405 0.21302232 0.08047341 0.06297287
#> [22,] 0.68849244 0.8428787 0.92122254 0.91559093 0.85499622 0.75434812
#> [23,] 0.74776050 0.6085080 0.53979615 0.47629695 0.41286357 0.41656941
#> [24,] 0.61410700 0.7869914 0.85151489 0.87561029 0.72726185 0.62134786
#> [25,] 0.62677533 0.7470999 0.76333160 0.67768681 0.55436215 0.37813272
#>              V39         V4        V40        V41        V42       V43
#>  [1,] 0.45334517 0.05002197 0.47209776 0.49084169 0.44589296 0.4486334
#>  [2,] 0.33283685 0.07057165 0.31025250 0.54149103 0.55064744 0.4559072
#>  [3,] 0.76269488 0.15613019 0.57670982 0.43617010 0.46687617 0.4503022
#>  [4,] 0.58134474 0.03677362 0.46849980 0.28212063 0.44342668 0.4258520
#>  [5,] 0.27831319 0.11902888 0.32761795 0.31873701 0.33770139 0.2365233
#>  [6,] 0.23821576 0.06347312 0.29090669 0.38341555 0.39932359 0.3382948
#>  [7,] 0.24239025 0.08166899 0.13840622 0.15033829 0.18975508 0.1743787
#>  [8,] 0.16462268 0.03579641 0.08862688 0.11107878 0.11833653 0.1693535
#>  [9,] 0.30673682 0.07188968 0.25860060 0.20156124 0.23757231 0.2487897
#> [10,] 0.41085145 0.03481520 0.41979671 0.26002754 0.16223215 0.3442786
#> [11,] 0.68721956 0.07794794 0.49934394 0.31540110 0.21223870 0.1853552
#> [12,] 0.12475613 0.02970361 0.13173036 0.05479147 0.08920428 0.1801954
#> [13,] 0.21983346 0.02926023 0.26555833 0.15312413 0.15201559 0.2170390
#> [14,] 0.29521560 0.06186398 0.28662879 0.28666914 0.29519927 0.2212975
#> [15,] 0.15908800 0.04006532 0.09628311 0.12175325 0.17863656 0.2714343
#> [16,] 0.23159301 0.02738182 0.15719433 0.11609634 0.15302857 0.1449768
#> [17,] 0.09399334 0.03240330 0.13372627 0.14479733 0.17128672 0.2620616
#> [18,] 0.13464523 0.03994206 0.11792627 0.07592035 0.04682438 0.0802058
#> [19,] 0.24266487 0.03685576 0.22183651 0.20812240 0.18519639 0.1872557
#> [20,] 0.23380000 0.01730000 0.23820000 0.33180000 0.38210000 0.1575000
#> [21,] 0.10834115 0.01983424 0.18269467 0.26101676 0.23999655 0.1825533
#> [22,] 0.73492936 0.04340732 0.69544969 0.58672656 0.49161337 0.3365403
#> [23,] 0.39673191 0.02835532 0.33093723 0.20198192 0.06918644 0.0927395
#> [24,] 0.59045385 0.01769627 0.59230406 0.47454002 0.43020879 0.3175812
#> [25,] 0.29902465 0.03725460 0.36961708 0.36642852 0.34258090 0.3044045
#>              V44        V45        V46        V47        V48        V49
#>  [1,] 0.52713339 0.49006617 0.34137148 0.24287699 0.19455986 0.10385677
#>  [2,] 0.41104242 0.58869591 0.57439969 0.39406524 0.24348597 0.13935998
#>  [3,] 0.39001662 0.49068458 0.39469655 0.08090084 0.13110756 0.06604621
#>  [4,] 0.46217625 0.51125948 0.27706717 0.15468871 0.17434386 0.10253702
#>  [5,] 0.15671273 0.17831338 0.20143669 0.19977110 0.15451356 0.07602332
#>  [6,] 0.12270862 0.11926018 0.06666463 0.08320201 0.09126958 0.05457666
#>  [7,] 0.06703610 0.07729988 0.04855977 0.04123890 0.03970721 0.02646428
#>  [8,] 0.14859703 0.09855774 0.08177569 0.12095992 0.11382203 0.06596499
#>  [9,] 0.19714546 0.12755093 0.09417936 0.06987965 0.04834546 0.03627812
#> [10,] 0.34383944 0.35074588 0.25308127 0.16444527 0.10312337 0.05812287
#> [11,] 0.21823238 0.39129478 0.28002448 0.14556298 0.14610513 0.09369308
#> [12,] 0.20071707 0.17865613 0.19252059 0.15263339 0.09456812 0.04543720
#> [13,] 0.23167187 0.12151308 0.09962381 0.10707328 0.09485752 0.07665123
#> [14,] 0.16989165 0.14818451 0.11335069 0.09616887 0.07603895 0.04713686
#> [15,] 0.17525087 0.02812701 0.02846442 0.03470432 0.03980162 0.01269639
#> [16,] 0.11428710 0.08091024 0.05657186 0.06529120 0.05138879 0.02365395
#> [17,] 0.36077989 0.34030321 0.22287049 0.10616001 0.04651817 0.02913633
#> [18,] 0.10345779 0.08489814 0.02341987 0.04573017 0.03832839 0.02071586
#> [19,] 0.11407663 0.07662680 0.08803599 0.10242260 0.06670105 0.03266467
#> [20,] 0.22280000 0.15820000 0.14330000 0.16340000 0.11330000 0.05670000
#> [21,] 0.09573007 0.08681652 0.08451334 0.04035126 0.03908811 0.03195631
#> [22,] 0.19104555 0.20209220 0.18669485 0.15660034 0.12617651 0.06791417
#> [23,] 0.12578165 0.14579933 0.07094747 0.03948657 0.04603431 0.02004269
#> [24,] 0.18860433 0.12814110 0.06472816 0.09331736 0.09591932 0.05089371
#> [25,] 0.18727991 0.10145234 0.07996925 0.07844524 0.05731770 0.02979327
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.08522675 0.028686399 0.016297125 0.011286533 0.011259284 0.008127532
#>  [2,] 0.07078781 0.048386783 0.035401791 0.025943860 0.017455330 0.016691993
#>  [3,] 0.13174890 0.039601904 0.018072066 0.033437464 0.034899165 0.024488265
#>  [4,] 0.09681703 0.028320682 0.031320977 0.017372809 0.008036025 0.009896857
#>  [5,] 0.12136212 0.018818218 0.021605522 0.020593031 0.014010829 0.010852662
#>  [6,] 0.10296923 0.007751231 0.012453578 0.014502671 0.006995814 0.010410978
#>  [7,] 0.09353666 0.017498131 0.019714672 0.018999273 0.018553050 0.018448588
#>  [8,] 0.04013186 0.022614220 0.012092561 0.009963187 0.007522127 0.013052535
#>  [9,] 0.13423854 0.010924937 0.007260862 0.013528849 0.010068654 0.011534561
#> [10,] 0.03192130 0.020108876 0.013468047 0.006507819 0.011714928 0.013938231
#> [11,] 0.05953481 0.038023172 0.025880767 0.012784810 0.010674437 0.030064112
#> [12,] 0.04423128 0.012663067 0.017294020 0.012939421 0.007240065 0.006907419
#> [13,] 0.05501928 0.018089317 0.029525832 0.013934258 0.013841558 0.019035346
#> [14,] 0.09847185 0.019886525 0.016163724 0.009866480 0.015382345 0.012229014
#> [15,] 0.02332345 0.017820976 0.018850497 0.016238811 0.015543397 0.005062425
#> [16,] 0.04017159 0.012687916 0.008750149 0.010538628 0.007687855 0.006365777
#> [17,] 0.03462907 0.018203230 0.012052907 0.004364825 0.005433237 0.011105814
#> [18,] 0.07586176 0.011866690 0.012218334 0.010684232 0.006572012 0.010650619
#> [19,] 0.04155289 0.031141866 0.020980589 0.019711106 0.013365571 0.010454801
#> [20,] 0.04760000 0.013300000 0.017000000 0.003500000 0.005200000 0.008300000
#> [21,] 0.04261753 0.010464925 0.008538168 0.011139468 0.004657176 0.004999169
#> [22,] 0.02970238 0.034513987 0.028120792 0.006877874 0.006673978 0.008189163
#> [23,] 0.03003590 0.008206250 0.005428723 0.006334973 0.010920213 0.006684708
#> [24,] 0.02504513 0.016434540 0.008338349 0.007805903 0.019788869 0.015182324
#> [25,] 0.04649855 0.019995690 0.015093759 0.009736306 0.011182025 0.009684731
#>               V55         V56         V57         V58         V59         V6
#>  [1,] 0.005719323 0.007101801 0.006552199 0.003725584 0.001553140 0.11753615
#>  [2,] 0.012529317 0.009418518 0.011115520 0.011379221 0.012634457 0.11062428
#>  [3,] 0.016318403 0.023000178 0.013219595 0.010358250 0.006044790 0.05746868
#>  [4,] 0.015563244 0.009646832 0.008787504 0.010390621 0.007289928 0.12023432
#>  [5,] 0.008655642 0.006846761 0.007196424 0.009582928 0.009000678 0.10964234
#>  [6,] 0.001091537 0.003051347 0.007279382 0.005932131 0.005202687 0.11279910
#>  [7,] 0.022202739 0.012430349 0.005971803 0.017501767 0.010169821 0.09642709
#>  [8,] 0.004826458 0.006985566 0.006158943 0.008511554 0.009975216 0.12403667
#>  [9,] 0.013316281 0.016051151 0.009207952 0.012299434 0.014947283 0.11547636
#> [10,] 0.011418088 0.009342575 0.006168535 0.009958963 0.008397962 0.06903885
#> [11,] 0.019921986 0.020285989 0.013432196 0.017905287 0.019354972 0.11959384
#> [12,] 0.004322043 0.003439851 0.003553008 0.002610142 0.003428648 0.07502574
#> [13,] 0.019223738 0.006225944 0.004111979 0.013934095 0.010898950 0.14481316
#> [14,] 0.008071881 0.010851256 0.005925703 0.006308450 0.005815954 0.13385489
#> [15,] 0.004577114 0.006028138 0.005467601 0.005366707 0.016577133 0.06148281
#> [16,] 0.006339971 0.005865402 0.005086787 0.006337579 0.005106928 0.07261297
#> [17,] 0.002491601 0.002299354 0.005079684 0.004075808 0.006774515 0.04499434
#> [18,] 0.005965211 0.008153813 0.006618051 0.004037082 0.006273219 0.11386238
#> [19,] 0.010461437 0.003031475 0.007774384 0.006434553 0.006380836 0.05236414
#> [20,] 0.007800000 0.007500000 0.010500000 0.016000000 0.009500000 0.08160000
#> [21,] 0.006309744 0.002357449 0.003212351 0.002832941 0.001867029 0.07692006
#> [22,] 0.014156450 0.011164048 0.004535588 0.005561562 0.007330326 0.08341415
#> [23,] 0.006718351 0.008548005 0.005106250 0.001206250 0.005396410 0.03175505
#> [24,] 0.011154109 0.006293950 0.007838027 0.012925735 0.019256170 0.03740933
#> [25,] 0.007297528 0.009239971 0.014555860 0.006408377 0.005711236 0.09901370
#>               V60         V7         V8          V9
#>  [1,] 0.002850412 0.15364750 0.13820837 0.284591284
#>  [2,] 0.009178225 0.10468290 0.12886176 0.230186619
#>  [3,] 0.012092496 0.10436940 0.30449900 0.593368204
#>  [4,] 0.005404361 0.12261163 0.15141132 0.256478649
#>  [5,] 0.006804531 0.11861327 0.07841764 0.121243045
#>  [6,] 0.006245314 0.05411524 0.07027780 0.224054212
#>  [7,] 0.007559874 0.12881702 0.12205188 0.143796301
#>  [8,] 0.008431661 0.15153028 0.14208404 0.191356831
#>  [9,] 0.007283776 0.11524610 0.09450646 0.090781953
#> [10,] 0.005357941 0.07785826 0.13765505 0.291109351
#> [11,] 0.008454193 0.17250935 0.43659323 0.588519964
#> [12,] 0.004485975 0.12220598 0.12934304 0.180278900
#> [13,] 0.010957647 0.19083632 0.18945205 0.224442562
#> [14,] 0.003832766 0.14167364 0.12569816 0.172888773
#> [15,] 0.009793305 0.03745423 0.10065814 0.170064371
#> [16,] 0.004679091 0.09489542 0.09367878 0.114683744
#> [17,] 0.001920316 0.09245384 0.08406108 0.086260303
#> [18,] 0.004419671 0.12031066 0.08519680 0.124160940
#> [19,] 0.005045154 0.06908306 0.09049955 0.111766981
#> [20,] 0.001100000 0.09930000 0.03150000 0.073600000
#> [21,] 0.001983162 0.08632751 0.07882307 0.112246062
#> [22,] 0.006818841 0.10047888 0.10758947 0.008010875
#> [23,] 0.003387899 0.04102686 0.02305439 0.034266756
#> [24,] 0.005919084 0.03557092 0.06309250 0.053809032
#> [25,] 0.005199137 0.12621670 0.14180420 0.102274300
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696