Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11         V12        V13        V14
#>  [1,] 0.021167229  0.23255275 0.16857169 0.075208501 0.12344431 0.27851749
#>  [2,] 0.017868625  0.28655139 0.32239745 0.297051387 0.29927703 0.25432135
#>  [3,] 0.036355502  0.28371582 0.37361058 0.469486017 0.52574422 0.53685213
#>  [4,] 0.025148313  0.22475872 0.27358456 0.311808474 0.33800932 0.38698448
#>  [5,] 0.116963132  0.36701828 0.48626228 0.557479688 0.67900249 0.79149035
#>  [6,] 0.050037346  0.20751871 0.25269013 0.269590658 0.32212614 0.29399760
#>  [7,] 0.029712868  0.18734602 0.27885026 0.333097201 0.27327094 0.24827554
#>  [8,] 0.020107476  0.29964912 0.32910452 0.384486129 0.35159615 0.27763037
#>  [9,] 0.053706888 -0.07562495 0.10563059 0.264795183 0.34846742 0.41782517
#> [10,] 0.038846660  0.31379179 0.35857180 0.372555163 0.27285982 0.11664941
#> [11,] 0.067430825  0.48413012 0.39714560 0.251626935 0.20897460 0.22976801
#> [12,] 0.020580308  0.19532033 0.18556292 0.153338508 0.14005905 0.11290900
#> [13,] 0.039606477  0.40763367 0.50280560 0.552396773 0.51982166 0.40080831
#> [14,] 0.010267004  0.11725507 0.10420351 0.007374607 0.04462095 0.04566230
#> [15,] 0.006916675  0.18138936 0.12068443 0.159637908 0.25756805 0.01592437
#> [16,] 0.029886757  0.21791276 0.26811607 0.372890649 0.45269547 0.48464358
#> [17,] 0.011250914  0.07529391 0.04753631 0.143550532 0.14489861 0.10610036
#> [18,] 0.012600000  0.42840000 0.30150000 0.120700000 0.32990000 0.57070000
#> [19,] 0.028834343  0.08759295 0.24597958 0.364938079 0.28222681 0.28187929
#> [20,] 0.015841507  0.11373509 0.11976375 0.100406896 0.12053025 0.14423501
#> [21,] 0.017572319  0.10593051 0.10388113 0.118096556 0.13327392 0.15654338
#> [22,] 0.008558281  0.10828543 0.10271193 0.082727815 0.08012316 0.08934768
#> [23,] 0.029300000  0.08600000 0.04140000 0.047200000 0.08350000 0.09380000
#> [24,] 0.027754048  0.31769646 0.24543978 0.080290997 0.18809535 0.46319137
#> [25,] 0.017392208  0.09194765 0.07361215 0.099178775 0.23538466 0.29975162
#>                V15        V16        V17       V18       V19         V2
#>  [1,]  0.422112888 0.56514249 0.61495887 0.5741432 0.6820294 0.03407336
#>  [2,]  0.160514535 0.12179117 0.27396193 0.4470423 0.6604343 0.04417890
#>  [3,]  0.611968817 0.73024879 0.85533027 0.9318248 0.9703523 0.05532900
#>  [4,]  0.399768574 0.45621496 0.56309065 0.6551238 0.7055023 0.02868315
#>  [5,]  0.997383175 1.04846635 0.85250235 0.8788452 0.9363184 0.13043782
#>  [6,]  0.287867452 0.30437656 0.31097817 0.2012698 0.2850053 0.06934609
#>  [7,]  0.243631805 0.23740239 0.22034191 0.2322260 0.3095948 0.05007009
#>  [8,]  0.202262896 0.22919592 0.22715055 0.3361036 0.5486911 0.02286702
#>  [9,]  0.542141174 0.55611624 0.75996608 0.7692344 0.8033375 0.06367156
#> [10,]  0.091213084 0.17984329 0.18483598 0.2240504 0.2439858 0.02211595
#> [11,]  0.339561503 0.39506491 0.23055346 0.2150111 0.2964849 0.07021580
#> [12,]  0.262189643 0.40579540 0.50250930 0.5164297 0.5938714 0.02832146
#> [13,]  0.293599144 0.34391506 0.36813109 0.3470138 0.4454027 0.05214804
#> [14,]  0.105590585 0.06071241 0.09582621 0.2061773 0.3418737 0.01444795
#> [15,] -0.004754037 0.20659104 0.35513189 0.3136307 0.5665822 0.02838262
#> [16,]  0.579138061 0.65201964 0.60333397 0.5030965 0.3611890 0.04406389
#> [17,]  0.108586546 0.14760543 0.14762215 0.2188801 0.1768190 0.02120325
#> [18,]  0.696200000 0.97510000 1.00000000 0.9293000 0.6210000 0.05190000
#> [19,]  0.319441688 0.43836126 0.57168541 0.4701215 0.4438583 0.05987016
#> [20,]  0.159746720 0.18988415 0.25588616 0.2974124 0.2911510 0.01587259
#> [21,]  0.170267388 0.18806027 0.15831553 0.1656811 0.1398915 0.01663937
#> [22,]  0.149450994 0.15909006 0.12118145 0.1526702 0.1622365 0.01498146
#> [23,]  0.146600000 0.08090000 0.11790000 0.2179000 0.3326000 0.06440000
#> [24,]  0.570103750 0.65089482 0.71584937 0.6410024 0.4517661 0.07210614
#> [25,]  0.278875061 0.27100890 0.18990032 0.2945728 0.6253456 0.03063929
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.8557679 0.9902177 1.0064204 0.8438921 0.5699879 0.3666761 0.4500011
#>  [2,] 0.7668004 0.6943754 0.4355446 0.2402184 0.2873792 0.4968280 0.8264267
#>  [3,] 0.9220654 0.7936517 0.6200336 0.4966249 0.3651235 0.2551442 0.1984008
#>  [4,] 0.7806981 0.8317819 0.8964253 0.9527682 0.9723201 0.9647904 0.9364189
#>  [5,] 1.0093710 0.8968719 0.8134396 0.9411037 0.8989059 0.6221542 0.5246040
#>  [6,] 0.4157028 0.4792436 0.5216748 0.5635655 0.5431784 0.5601154 0.5835148
#>  [7,] 0.2778197 0.4087423 0.4885620 0.5778665 0.6825306 0.6824581 0.6885789
#>  [8,] 0.6787670 0.8091665 0.8043765 0.7928385 0.9041952 0.9090348 0.9314829
#>  [9,] 0.9386858 0.9131758 0.8885736 0.9502127 1.0589625 0.9073919 0.5934995
#> [10,] 0.3184206 0.4373401 0.6289567 0.7639764 0.8155770 0.9349527 1.0243228
#> [11,] 0.5736818 0.5458414 0.2885823 0.2210413 0.3179070 0.3977058 0.5548713
#> [12,] 0.7814049 0.9309149 1.0084713 0.9000765 0.7302083 0.5358602 0.2974561
#> [13,] 0.6379584 0.7529011 0.8578658 0.8485380 0.8734760 0.8981169 0.8867717
#> [14,] 0.3624395 0.4529747 0.8294337 0.9238098 0.8694941 0.7123853 0.7289434
#> [15,] 0.3894632 0.4979310 0.4729369 0.4241691 0.6171462 0.7671582 0.6508893
#> [16,] 0.4502663 0.5664545 0.6140488 0.5670147 0.6690450 0.7376615 0.6461688
#> [17,] 0.1574174 0.2066552 0.1319436 0.1362040 0.1867455 0.1124504 0.2225954
#> [18,] 0.4586000 0.5001000 0.5032000 0.7082000 0.8420000 0.8109000 0.7690000
#> [19,] 0.6337000 0.3950748 0.2099085 0.3634399 0.9079975 0.9334164 0.4825803
#> [20,] 0.3659525 0.4757192 0.6011143 0.7159213 0.7583748 0.7730117 0.8374726
#> [21,] 0.1447868 0.2393138 0.2677957 0.4135280 0.6333643 0.8945607 0.9990320
#> [22,] 0.1831444 0.2605219 0.3659187 0.5140584 0.6888120 0.8412047 0.9034834
#> [23,] 0.3258000 0.2111000 0.2302000 0.3361000 0.4259000 0.4609000 0.2606000
#> [24,] 0.3868887 0.5443424 0.7590896 0.8448827 0.8539880 0.7269412 0.6721946
#> [25,] 0.8553941 0.8070720 0.7601882 0.4883617 0.4462907 0.6480921 0.8025529
#>             V27        V28        V29          V3       V30        V31
#>  [1,] 0.5098312 0.41938358 0.32628623 0.020148042 0.4270765 0.19174519
#>  [2,] 1.0078985 0.92492951 0.85587811 0.046601867 0.7709311 0.57281014
#>  [3,] 0.1652165 0.08529838 0.08342104 0.063834202 0.1882836 0.20477152
#>  [4,] 0.9193761 0.91311004 0.82358892 0.028885181 0.6205759 0.39012688
#>  [5,] 0.5103287 0.51781842 0.46707858 0.087698387 0.3499076 0.29208138
#>  [6,] 0.6291724 0.77361118 0.79230719 0.080162399 0.8161712 0.85654039
#>  [7,] 0.7687563 0.80825719 0.79598377 0.064777249 0.6533588 0.59197554
#>  [8,] 0.9608439 0.97851773 0.89061808 0.038411518 0.6669377 0.45892779
#>  [9,] 0.3599125 0.23041248 0.30684335 0.127895973 0.3166738 0.29987498
#> [10,] 0.9897405 0.94677682 0.90227362 0.030815273 0.8250229 0.73467731
#> [11,] 0.5579455 0.71808902 0.70750051 0.045128273 0.8377611 0.63856897
#> [12,] 0.1066901 0.02278395 0.23890569 0.012540628 0.3654475 0.31389609
#> [13,] 0.9085270 0.77400196 0.56923070 0.048546955 0.4284991 0.39998178
#> [14,] 0.6204160 0.69431974 0.81642869 0.037929085 0.7864465 0.61081280
#> [15,] 0.7335620 0.83053115 0.67149447 0.047313343 0.2892519 0.04999547
#> [16,] 0.6005237 0.74875718 0.81405356 0.039529021 0.6530275 0.52265728
#> [17,] 0.3850649 0.61119035 0.69196539 0.025808968 0.6454041 0.61236896
#> [18,] 0.8105000 0.62030000 0.23560000 0.062100000 0.2595000 0.62990000
#> [19,] 0.3084203 0.67443294 0.80916392 0.062960344 0.4406085 0.52905441
#> [20,] 0.8593892 0.83560427 0.82781105 0.023562467 0.7612713 0.71198135
#> [21,] 0.8153497 0.61952544 0.52953884 0.012289840 0.3714724 0.28265329
#> [22,] 0.7884476 0.61572852 0.49999409 0.007384107 0.3604086 0.25820859
#> [23,] 0.0874000 0.28620000 0.56060000 0.039000000 0.8344000 0.80960000
#> [24,] 0.6527266 0.51479004 0.34953353 0.099132538 0.3934209 0.43895284
#> [25,] 0.9099212 0.78835717 0.56631223 0.021507385 0.4605175 0.38144346
#>             V32        V33        V34         V35        V36        V37
#>  [1,] 0.3344689 0.42214720 0.28884515  0.22442920 0.29762676 0.31102119
#>  [2,] 0.5192323 0.69034982 0.62762511  0.50671696 0.35865714 0.37720386
#>  [3,] 0.2668671 0.31756142 0.37630277  0.39001793 0.21394675 0.32636160
#>  [4,] 0.2727943 0.15654654 0.08685996  0.18082222 0.24105019 0.24041015
#>  [5,] 0.2397921 0.06279919 0.10798801  0.14596120 0.11457161 0.16438776
#>  [6,] 0.7595786 0.56979504 0.51555781  0.48939291 0.31390466 0.23599678
#>  [7,] 0.5111954 0.41525555 0.35629297  0.18469770 0.14650325 0.23209178
#>  [8,] 0.3232351 0.29827796 0.22049493  0.12015189 0.02680564 0.06340302
#>  [9,] 0.2223051 0.30944227 0.41183596  0.48985145 0.34777093 0.31847962
#> [10,] 0.4935870 0.40173544 0.27038140  0.11888334 0.02046509 0.10234494
#> [11,] 0.6187404 0.65963566 0.69114275  0.78261840 0.79403599 0.80301916
#> [12,] 0.5022263 0.40682199 0.16395855 -0.02259355 0.21642963 0.40528839
#> [13,] 0.4207888 0.40676905 0.30300245  0.09227506 0.31198405 0.38878717
#> [14,] 0.6489356 0.81660553 0.90795331  0.87027622 0.67402340 0.40818763
#> [15,] 0.2754298 0.58530184 0.85688719  0.90483369 0.87049650 0.60324246
#> [16,] 0.4411546 0.25976469 0.26025614  0.30766560 0.27804502 0.21132962
#> [17,] 0.5112489 0.62894471 0.73877582  0.71828925 0.76043483 0.74498988
#> [18,] 0.6762000 0.29030000 0.43930000  0.85290000 0.71800000 0.48010000
#> [19,] 0.6949658 0.58746697 0.88835358  0.79598483 0.72643765 0.76999778
#> [20,] 0.6263177 0.58484412 0.60316475  0.60298986 0.56977123 0.47943256
#> [21,] 0.1202008 0.14490770 0.17045170  0.21612540 0.18654335 0.04939869
#> [22,] 0.1748782 0.20880202 0.23793248  0.22548608 0.18815962 0.13113454
#> [23,] 0.7250000 0.80480000 0.94350000  1.00000000 0.89600000 0.55160000
#> [24,] 0.2995337 0.24756396 0.14105576  0.37131736 0.61553841 0.65702316
#> [25,] 0.2279054 0.28987999 0.31392334  0.27902646 0.26829801 0.37256161
#>              V38       V39          V4        V40         V41        V42
#>  [1,] 0.30501485 0.1708126 0.038114073 0.12063643  0.15009772 0.24930426
#>  [2,] 0.37658490 0.3326701 0.069183176 0.27397903  0.36319946 0.38539787
#>  [3,] 0.36820833 0.1908104 0.083557062 0.25597839  0.41873128 0.35592693
#>  [4,] 0.27856265 0.2938358 0.056744116 0.32717124  0.33079393 0.33215379
#>  [5,] 0.13700337 0.1793442 0.007026776 0.24401818  0.17630271 0.16169106
#>  [6,] 0.31188469 0.3441340 0.080443830 0.32966144  0.56659636 0.60020665
#>  [7,] 0.17359740 0.2061385 0.081829129 0.33848377  0.38649744 0.35494649
#>  [8,] 0.07491791 0.1409916 0.036864430 0.18217989  0.06998774 0.13946058
#>  [9,] 0.45352934 0.4124781 0.099002033 0.33770210  0.24623321 0.29150394
#> [10,] 0.13928993 0.1253113 0.044915428 0.06100997  0.10252601 0.13872967
#> [11,] 0.78246575 0.6613169 0.069608618 0.52580989  0.37138775 0.30442243
#> [12,] 0.36811835 0.2683384 0.045482248 0.16498789  0.15880331 0.08275687
#> [13,] 0.36219139 0.3883015 0.040731056 0.31161828  0.19710277 0.27480783
#> [14,] 0.22090547 0.2202904 0.031674847 0.32163010  0.29715157 0.21888674
#> [15,] 0.53201377 0.4088743 0.016402816 0.18134362 -0.04473050 0.33771719
#> [16,] 0.18711003 0.1720371 0.042975398 0.15941756  0.20269017 0.15782525
#> [17,] 0.67991801 0.6656286 0.038737909 0.61261499  0.50043475 0.41549431
#> [18,] 0.58560000 0.4993000 0.051800000 0.28660000  0.06010000 0.11670000
#> [19,] 0.45888085 0.3178390 0.089358476 0.32408649  0.28140874 0.28448182
#> [20,] 0.36536769 0.2642605 0.031179029 0.24116061  0.20779406 0.17007608
#> [21,] 0.10259408 0.1665024 0.023764237 0.23950788  0.30574094 0.28543902
#> [22,] 0.12468287 0.2145729 0.010800000 0.25504637  0.28999999 0.27826023
#> [23,] 0.30370000 0.2338000 0.017300000 0.23820000  0.33180000 0.38210000
#> [24,] 0.36194379 0.2605736 0.092436181 0.44939026  0.47972379 0.37575785
#> [25,] 0.45764130 0.4410448 0.027662921 0.29894391  0.19191933 0.15191674
#>              V43        V44        V45        V46        V47        V48
#>  [1,] 0.16821023 0.14198641 0.14023366 0.09289267 0.09029856 0.08404567
#>  [2,] 0.34742123 0.20939813 0.10136969 0.08079653 0.06906342 0.05225073
#>  [3,] 0.16773173 0.09028062 0.07307312 0.07955326 0.06401908 0.04355051
#>  [4,] 0.28761247 0.21014451 0.15031642 0.13208854 0.13437910 0.09499668
#>  [5,] 0.16535180 0.25828696 0.13129353 0.06400103 0.08100526 0.08397413
#>  [6,] 0.44859585 0.39506513 0.57029320 0.54382829 0.34198701 0.19310637
#>  [7,] 0.23232835 0.12572249 0.17127840 0.21136373 0.21887776 0.19161214
#>  [8,] 0.24821564 0.29959094 0.29825790 0.25588087 0.19048747 0.12970516
#>  [9,] 0.24756805 0.22024524 0.24955366 0.20766850 0.09858226 0.04508138
#> [10,] 0.20197972 0.16879526 0.09738496 0.07848149 0.10549047 0.09994033
#> [11,] 0.31564189 0.40207451 0.48381747 0.34301708 0.17652912 0.15026164
#> [12,] 0.03129153 0.06860199 0.04428016 0.03276119 0.03674976 0.03317549
#> [13,] 0.42771394 0.44000782 0.39429378 0.19287123 0.10233838 0.10848734
#> [14,] 0.30761467 0.23519267 0.12390916 0.06449108 0.04163555 0.03362951
#> [15,] 0.30404574 0.46651174 0.19343820 0.04874466 0.05826614 0.13371565
#> [16,] 0.17448905 0.15615692 0.11696809 0.13574547 0.10369384 0.04271988
#> [17,] 0.28849808 0.15526541 0.13668795 0.12468221 0.12378867 0.10259271
#> [18,] 0.27370000 0.28120000 0.20780000 0.06600000 0.04910000 0.03450000
#> [19,] 0.05547342 0.18120753 0.15541776 0.16473671 0.12451226 0.07454092
#> [20,] 0.17988885 0.12513539 0.07162650 0.06647145 0.04802046 0.03908063
#> [21,] 0.19250278 0.07580015 0.07955906 0.09260133 0.04352949 0.02780548
#> [22,] 0.20630527 0.15659471 0.13100533 0.13279207 0.06886624 0.04228412
#> [23,] 0.15750000 0.22280000 0.15820000 0.14330000 0.16340000 0.11330000
#> [24,] 0.22495682 0.16403416 0.16942204 0.13831968 0.13992786 0.13607685
#> [25,] 0.07995194 0.03518436 0.08848242 0.09116064 0.06722035 0.03624498
#>               V49         V5           V50          V51           V52
#>  [1,] 0.061785333 0.04379314  0.0269304989  0.013726259  0.0117015878
#>  [2,] 0.024277838 0.16252697  0.0117176235  0.012342590  0.0120237646
#>  [3,] 0.038576863 0.08726923  0.0276550732  0.015753991  0.0343919457
#>  [4,] 0.061091954 0.12825015  0.0260309311  0.020923622  0.0139528942
#>  [5,] 0.070560259 0.05405544  0.0099310629  0.030450821  0.0185595751
#>  [6,] 0.108136171 0.07824296  0.0419413368  0.032131897  0.0293220940
#>  [7,] 0.094375652 0.08337230  0.0138965621  0.021792340  0.0243725172
#>  [8,] 0.064582595 0.07679881  0.0209054495  0.021625874  0.0149616265
#>  [9,] 0.039807699 0.18104190 -0.0005113275 -0.004845868  0.0118335391
#> [10,] 0.063396974 0.04382407  0.0160324309  0.011602909  0.0097178228
#> [11,] 0.083380278 0.07468401  0.0320614736  0.020228878  0.0214131210
#> [12,] 0.007001495 0.04453459  0.0067942993  0.012370492  0.0080755610
#> [13,] 0.062687233 0.03681259  0.0144859638  0.021191690  0.0106054900
#> [14,] 0.010898768 0.06012781  0.0138550765  0.013455371  0.0141881302
#> [15,] 0.018924237 0.12218980  0.0334033805  0.026647681 -0.0007218234
#> [16,] 0.028755758 0.06790425  0.0201240491  0.013007857  0.0096316566
#> [17,] 0.055333195 0.03801294  0.0243688179  0.017017892  0.0086609868
#> [18,] 0.017200000 0.10720000  0.0287000000  0.002700000  0.0208000000
#> [19,] 0.043899896 0.15532922  0.0070641837  0.016130131  0.0039526046
#> [20,] 0.016811312 0.04531424  0.0122500424  0.007393396  0.0065945935
#> [21,] 0.027709357 0.05866264  0.0088473472  0.012685548  0.0081358350
#> [22,] 0.032300669 0.02164437  0.0102781461  0.002327815  0.0067052977
#> [23,] 0.056700000 0.04760000  0.0133000000  0.017000000  0.0035000000
#> [24,] 0.078331230 0.09541970  0.0161131545  0.024168788  0.0145682042
#> [25,] 0.011435989 0.04231214  0.0226137262  0.018852323  0.0139134177
#>                V53         V54         V55         V56         V57         V58
#>  [1,]  0.002490576 0.010435419 0.002452950 0.010531424 0.004323248 0.009989271
#>  [2,]  0.011316689 0.009804806 0.002454873 0.002149131 0.002135379 0.004817356
#>  [3,]  0.020366487 0.013786018 0.025393886 0.012584719 0.001424450 0.017306961
#>  [4,]  0.018260544 0.014299494 0.007198282 0.009431434 0.008530194 0.013822360
#>  [5,]  0.016462223 0.018783051 0.019576477 0.004218266 0.002922505 0.012938230
#>  [6,]  0.018666884 0.015079219 0.014604308 0.009444409 0.010886359 0.010862350
#>  [7,]  0.011190491 0.012605219 0.011450924 0.007577123 0.008614010 0.009183782
#>  [8,]  0.008434289 0.005838548 0.002228789 0.003504968 0.004899430 0.001396278
#>  [9,]  0.018277802 0.008501942 0.005417646 0.015844771 0.013145711 0.014145679
#> [10,]  0.007757910 0.007606465 0.004831472 0.005226456 0.006042149 0.005947152
#> [11,]  0.011089156 0.014054384 0.012505607 0.008691982 0.010368236 0.010045079
#> [12,] -0.001091843 0.003928496 0.003211523 0.002696866 0.007972741 0.004194952
#> [13,]  0.006122138 0.011205281 0.011594189 0.007147009 0.007308576 0.008527867
#> [14,]  0.018843360 0.014344236 0.005878766 0.012217722 0.019978259 0.007676652
#> [15,]  0.006593152 0.018431999 0.005777776 0.020965335 0.022620943 0.008710616
#> [16,]  0.009660573 0.007932914 0.009887990 0.010999299 0.009298363 0.006041988
#> [17,]  0.015197678 0.012141784 0.012786799 0.009620342 0.006365678 0.007906156
#> [18,]  0.004800000 0.019900000 0.012600000 0.002200000 0.003700000 0.003400000
#> [19,]  0.015007948 0.007029342 0.020606474 0.007675368 0.014813737 0.005215605
#> [20,]  0.006083647 0.006022764 0.007028465 0.005258946 0.006482817 0.005045979
#> [21,]  0.006069593 0.007196433 0.007761909 0.003714055 0.002628514 0.002442629
#> [22,]  0.003807287 0.005092056 0.004087416 0.003210595 0.002572185 0.004239736
#> [23,]  0.005200000 0.008300000 0.007800000 0.007500000 0.010500000 0.016000000
#> [24,]  0.011050017 0.012518576 0.010166372 0.006543884 0.010031097 0.009699608
#> [25,]  0.023521755 0.022416614 0.010303457 0.011300299 0.004234741 0.009041651
#>               V59         V6          V60         V7         V8         V9
#>  [1,] 0.002789196 0.03641012 0.0067884876 0.06716972 0.11397466 0.24315984
#>  [2,] 0.007264753 0.08646822 0.0021102802 0.05666007 0.10277422 0.31361801
#>  [3,] 0.016062692 0.06079728 0.0093874697 0.08238907 0.03690385 0.12400944
#>  [4,] 0.008864427 0.14916282 0.0063355796 0.13893841 0.14918504 0.18476523
#>  [5,] 0.010337070 0.16237368 0.0151834660 0.17274451 0.15190507 0.24417059
#>  [6,] 0.013411728 0.12702802 0.0094683834 0.09906547 0.13415130 0.21080601
#>  [7,] 0.008468040 0.12718204 0.0042452976 0.13892846 0.12022536 0.12596275
#>  [8,] 0.001868997 0.10857615 0.0041116601 0.10228342 0.14883470 0.22570963
#>  [9,] 0.015836175 0.18486482 0.0042667718 0.17775137 0.03457185 0.04603812
#> [10,] 0.006463879 0.09647867 0.0069294131 0.12505728 0.17152102 0.22353432
#> [11,] 0.007144056 0.11855187 0.0075139924 0.19142518 0.30087125 0.47621306
#> [12,] 0.001466954 0.07936123 0.0002503444 0.15351505 0.14722062 0.20438296
#> [13,] 0.005466827 0.10084357 0.0044746242 0.08792812 0.18330766 0.32241124
#> [14,] 0.007380676 0.05698258 0.0118900578 0.09166160 0.07518185 0.02261804
#> [15,] 0.009866495 0.11145906 0.0020482307 0.17380721 0.16332065 0.34414119
#> [16,] 0.005599859 0.11520353 0.0049249594 0.14165022 0.16445509 0.20615215
#> [17,] 0.011764612 0.07286926 0.0083168502 0.06376358 0.07698674 0.02242501
#> [18,] 0.011400000 0.25870000 0.0077000000 0.23040000 0.20670000 0.34160000
#> [19,] 0.019340527 0.22411479 0.0074742893 0.21418016 0.19672882 0.14251313
#> [20,] 0.005490585 0.08627121 0.0050252688 0.08607310 0.09604748 0.10888304
#> [21,] 0.004373969 0.07017313 0.0022142762 0.07147991 0.05942765 0.10322557
#> [22,] 0.001517882 0.02428344 0.0038615899 0.05637153 0.06804174 0.09794568
#> [23,] 0.009500000 0.08160000 0.0011000000 0.09930000 0.03150000 0.07360000
#> [24,] 0.010399040 0.11379079 0.0096341062 0.10637730 0.16890241 0.23575038
#> [25,] 0.001082865 0.02582803 0.0044403591 0.03828805 0.05951865 0.10871571
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768