Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.049938736 0.27516155 0.30967473 0.28513469 0.21599050 0.09830080
#>  [2,] 0.050390343 0.14225667 0.26673405 0.28226078 0.41278286 0.35953078
#>  [3,] 0.028011825 0.25200556 0.30267498 0.30702734 0.26228375 0.29121315
#>  [4,] 0.036295264 0.34914385 0.37504838 0.21795088 0.47605867 0.90495543
#>  [5,] 0.043288338 0.29525541 0.39036682 0.51411261 0.58782008 0.70338330
#>  [6,] 0.028618899 0.31774937 0.38584552 0.40350187 0.32453522 0.21262979
#>  [7,] 0.027076362 0.26583516 0.31788423 0.35655141 0.31099769 0.20772124
#>  [8,] 0.040436696 0.15619316 0.22416055 0.27695288 0.27563296 0.27829225
#>  [9,] 0.016428942 0.30450578 0.33370706 0.29785205 0.29577871 0.25037918
#> [10,] 0.032606877 0.30565747 0.32782071 0.27027064 0.25388943 0.21921207
#> [11,] 0.027015172 0.18243668 0.22530260 0.25174633 0.28830552 0.34544719
#> [12,] 0.158014584 0.09174491 0.33298339 0.48753818 0.69913168 0.70997564
#> [13,] 0.048584024 0.12468445 0.11031276 0.13709467 0.26251839 0.28765655
#> [14,] 0.004684631 0.10631169 0.12807608 0.19060442 0.31384387 0.27363602
#> [15,] 0.027617905 0.04800342 0.01895689 0.17084274 0.15698590 0.12085320
#> [16,] 0.026716879 0.11784584 0.23394292 0.35728454 0.26746226 0.24182120
#> [17,] 0.015743950 0.06530982 0.10012019 0.10287825 0.13587336 0.14525254
#> [18,] 0.021657003 0.10939308 0.12573368 0.11125405 0.12874506 0.11635627
#> [19,] 0.021486367 0.31209035 0.31073465 0.22118581 0.33105850 0.52838740
#> [20,] 0.016151455 0.10669860 0.12389964 0.09590422 0.09201113 0.13625437
#> [21,] 0.010570094 0.10457341 0.10029288 0.09854858 0.11211756 0.12519232
#> [22,] 0.029453140 0.01136683 0.08961606 0.22488861 0.16634527 0.09264279
#> [23,] 0.021486879 0.06222709 0.07267636 0.10888069 0.20052443 0.34564330
#> [24,] 0.065821079 0.27650651 0.24496298 0.09715823 0.10564955 0.44168287
#> [25,] 0.013041118 0.14368000 0.12694722 0.09441940 0.07738552 0.11027895
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.08691319 0.15543392 0.1652408 0.2273645 0.1937357 0.039116876 0.1604829
#>  [2,] 0.21947323 0.26033792 0.3191393 0.1647956 0.2604803 0.083806418 0.3637670
#>  [3,] 0.24321909 0.21507523 0.1627882 0.1886933 0.3072103 0.027124965 0.3408082
#>  [4,] 0.93201018 0.86940814 0.7302658 0.8255646 0.9223096 0.047962015 0.8804958
#>  [5,] 0.77792814 0.84047714 0.9726476 0.9272033 0.9079249 0.047022780 0.8214513
#>  [6,] 0.17258036 0.18594543 0.1970786 0.2952867 0.5161779 0.044227160 0.6593851
#>  [7,] 0.17198662 0.24639096 0.2519449 0.3853097 0.6014490 0.027441858 0.7295707
#>  [8,] 0.23782840 0.14705752 0.1649795 0.2098167 0.3345303 0.061886464 0.3754750
#>  [9,] 0.14747213 0.11000417 0.2583714 0.4221190 0.6402987 0.043025068 0.7554987
#> [10,] 0.20665648 0.32147552 0.4066830 0.4269948 0.5347015 0.044758272 0.6156907
#> [11,] 0.40558829 0.49285539 0.5878198 0.6848185 0.7739450 0.032365144 0.8471054
#> [12,] 0.66077717 0.74498207 0.8791228 0.8690013 0.9128718 0.140242279 0.7800712
#> [13,] 0.32555119 0.33779777 0.3584872 0.2147502 0.2527008 0.067156481 0.3108322
#> [14,] 0.20041976 0.33459317 0.4315824 0.5452498 0.7467820 0.036041742 0.7707946
#> [15,] 0.14044763 0.18557735 0.1662505 0.1037399 0.1274524 0.047154831 0.1424691
#> [16,] 0.23824028 0.35234498 0.5710035 0.5368876 0.5064845 0.053522470 0.6649629
#> [17,] 0.17100597 0.30529824 0.4077881 0.3847613 0.3255567 0.017429773 0.3195358
#> [18,] 0.08307868 0.08678199 0.1208323 0.1917763 0.3781078 0.025288007 0.3602571
#> [19,] 0.64737447 0.83777747 0.9326080 0.9105134 0.6979311 0.029982748 0.4641793
#> [20,] 0.10269456 0.11795284 0.2029369 0.2545168 0.2734092 0.021235269 0.4269331
#> [21,] 0.17416794 0.17965245 0.1468538 0.1539578 0.1116651 0.014029774 0.1343508
#> [22,] 0.10075364 0.17664401 0.2006180 0.2480934 0.3100144 0.041126040 0.1700897
#> [23,] 0.51416812 0.69185499 0.7643188 0.7816984 0.8003362 0.021861162 0.8337946
#> [24,] 0.65617855 0.66120407 0.6814726 0.5436783 0.3503870 0.053896163 0.3024548
#> [25,] 0.09404787 0.12739641 0.2403054 0.2803975 0.3551256 0.005191853 0.5244475
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.3133866 0.58065693 0.7492112 0.8292475 0.95278819 1.0323649 0.9888967
#>  [2,] 0.5370039 0.69658864 0.7087131 0.6968731 0.54680618 0.5346557 0.6151405
#>  [3,] 0.3515416 0.32759213 0.3946153 0.4644660 0.50436039 0.6723969 0.8561866
#>  [4,] 0.8818772 0.92780905 0.9866733 0.9009781 0.75991165 0.7527063 0.6000974
#>  [5,] 0.7767543 0.72830595 0.5276615 0.2731151 0.22184809 0.1822627 0.2584056
#>  [6,] 0.7908682 0.84931034 0.7797770 0.8486322 0.86244639 0.9438706 0.9853428
#>  [7,] 0.8223523 0.77745168 0.8336767 0.9495031 0.91344950 0.9281171 0.9761973
#>  [8,] 0.4119837 0.38474087 0.4879322 0.6313957 0.64661776 0.7478211 0.8949796
#>  [9,] 0.6881718 0.45850450 0.2380460 0.2616231 0.50615509 0.8598520 1.0368320
#> [10,] 0.6473429 0.76401792 0.8614728 0.8206483 0.76622685 0.7069884 0.6008547
#> [11,] 0.9094726 0.93711635 0.9397667 0.9172809 0.86148491 0.8271164 0.8516355
#> [12,] 0.8186724 0.81005843 0.8713283 0.8432139 0.55380693 0.5687926 0.4406777
#> [13,] 0.3095364 0.40986634 0.4837938 0.3897347 0.42538882 0.4514983 0.4555681
#> [14,] 0.7368106 0.61614604 0.5684240 0.5853851 0.66545256 0.7429115 0.7414005
#> [15,] 0.2322149 0.09796317 0.1837621 0.2250157 0.09089141 0.2317439 0.5361953
#> [16,] 0.4811659 0.35447742 0.5252197 0.9428266 0.85753135 0.4203563 0.3015578
#> [17,] 0.4196880 0.55812886 0.6372561 0.6775121 0.74490380 0.8211305 0.7710097
#> [18,] 0.2275088 0.25425884 0.3995586 0.5479316 0.53841280 0.4757046 0.3819978
#> [19,] 0.4512931 0.57593764 0.6758025 0.6802917 0.61296578 0.6008042 0.6137390
#> [20,] 0.5150276 0.62930685 0.8334664 0.8335605 0.82059660 0.8709935 0.8702423
#> [21,] 0.2166997 0.24786522 0.3786646 0.5991426 0.83720112 0.9797198 0.8365840
#> [22,] 0.2771119 0.12512547 0.2255626 0.2518033 0.07280728 0.2317076 0.3808867
#> [23,] 0.9527680 0.91484840 0.8842038 0.9345102 0.93649633 0.9548483 0.8752680
#> [24,] 0.5743706 0.80718643 0.8780389 0.8679919 0.84522872 0.6435915 0.6097153
#> [25,] 0.6479230 0.80609523 0.9414751 0.9011214 0.88330302 0.9588180 0.9799560
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.9910857 0.8687436 0.022600089 0.7106769 0.7863986 0.7266811 0.51433627
#>  [2,] 0.8025728 0.8665482 0.072454834 0.8483456 0.9418950 0.9126141 0.71611766
#>  [3,] 0.9664399 1.0130749 0.011657608 0.8189144 0.5333823 0.4827151 0.52293235
#>  [4,] 0.4404862 0.5196598 0.028587292 0.3648629 0.2946900 0.1231663 0.02816775
#>  [5,] 0.2423403 0.1071343 0.036795690 0.1445222 0.0912021 0.1703657 0.26244624
#>  [6,] 0.9451334 0.8577257 0.039924897 0.6769497 0.5259156 0.4072759 0.31828302
#>  [7,] 1.0123890 0.8795120 0.038399512 0.6260840 0.3964008 0.2602691 0.18777336
#>  [8,] 0.9135185 0.8138618 0.072334792 0.6814979 0.6320088 0.4404164 0.28846497
#>  [9,] 0.9349205 0.8796364 0.043151446 0.7977004 0.5698697 0.5039047 0.71067313
#> [10,] 0.4696155 0.3679476 0.047283601 0.4983132 0.5351592 0.5460590 0.53562869
#> [11,] 0.8199367 0.6762244 0.033886336 0.5110815 0.3670872 0.2942185 0.25092941
#> [12,] 0.5012319 0.4082237 0.143534522 0.2233096 0.1207163 0.1509963 0.29877930
#> [13,] 0.7012572 0.7817552 0.084485999 0.7283454 0.7554390 0.7117897 0.68732819
#> [14,] 0.5290620 0.3163632 0.039445102 0.2046060 0.3512389 0.4819859 0.47651686
#> [15,] 0.7767802 0.7891862 0.026478667 0.7067251 0.7209761 0.5761245 0.66186997
#> [16,] 0.5820606 0.7206539 0.074201669 0.4532721 0.6010200 0.7352563 0.61511002
#> [17,] 0.7666550 0.7821898 0.023420880 0.7383680 0.6729313 0.6271051 0.67876851
#> [18,] 0.5939592 0.8847279 0.022416236 0.9630503 0.9223908 0.8487683 0.77597688
#> [19,] 0.5091980 0.3651104 0.045012289 0.4684534 0.4557031 0.3925728 0.25798502
#> [20,] 0.8103541 0.7281568 0.032510982 0.6238736 0.5411989 0.4760885 0.47157781
#> [21,] 0.6158127 0.5044065 0.011885068 0.3830278 0.2981651 0.1595090 0.17668475
#> [22,] 0.5569445 0.7205603 0.036614496 0.7688134 0.8132433 0.5716345 0.58616217
#> [23,] 0.6434194 0.3717639 0.019852157 0.3543167 0.4054514 0.3084614 0.27879896
#> [24,] 0.6407878 0.4455113 0.064326164 0.4995357 0.4885649 0.3042651 0.26587008
#> [25,] 0.9248708 0.8519891 0.009179269 0.7321115 0.6357810 0.5949673 0.56421506
#>             V34       V35        V36        V37        V38        V39
#>  [1,] 0.1954409 0.1138070 0.13062637 0.10830667 0.14597163 0.19882087
#>  [2,] 0.6842244 0.7326812 0.42529689 0.15839716 0.22770105 0.20583250
#>  [3,] 0.3822390 0.1636581 0.11740305 0.18473386 0.16123301 0.09492840
#>  [4,] 0.1287146 0.1953011 0.20373147 0.14497872 0.14938333 0.30399941
#>  [5,] 0.1833231 0.2771591 0.07121517 0.18831969 0.28603651 0.22596194
#>  [6,] 0.2466214 0.1378381 0.08679554 0.11683481 0.17068192 0.22979007
#>  [7,] 0.1361787 0.1160436 0.02324104 0.05900627 0.08236048 0.08575629
#>  [8,] 0.2536721 0.1653556 0.07102188 0.20454535 0.19201522 0.24258225
#>  [9,] 0.6696674 0.5261800 0.34472558 0.35815163 0.38398856 0.34669344
#> [10,] 0.5303742 0.3847437 0.27769994 0.23527444 0.27955159 0.40865649
#> [11,] 0.1545827 0.1689760 0.20627263 0.26760047 0.32310801 0.32468152
#> [12,] 0.3297012 0.3505030 0.18192038 0.20659578 0.34903539 0.38373372
#> [13,] 0.5253982 0.4455781 0.31479043 0.15313943 0.29023038 0.26344533
#> [14,] 0.4212085 0.4033519 0.51168055 0.59142289 0.37412308 0.38786270
#> [15,] 0.8479941 0.9115140 0.89392836 0.71417965 0.60023584 0.58053097
#> [16,] 0.8604576 0.7095238 0.65331252 0.70605459 0.47412345 0.35314752
#> [17,] 0.7422456 0.7771826 0.71088329 0.55777814 0.47878824 0.42265716
#> [18,] 0.7207334 0.6473588 0.55636407 0.48538384 0.21947842 0.10428790
#> [19,] 0.1819703 0.4098072 0.53580673 0.50138303 0.37784701 0.24792346
#> [20,] 0.5259770 0.5247480 0.48589223 0.37645939 0.25136833 0.14345352
#> [21,] 0.2102345 0.2584283 0.19387148 0.06957704 0.06466205 0.14719092
#> [22,] 0.8087057 0.9153298 0.91760610 0.76529623 0.75352074 0.78706231
#> [23,] 0.2504424 0.1603397 0.05328882 0.13094036 0.12439589 0.12748214
#> [24,] 0.1775059 0.4666642 0.69392252 0.73982434 0.42682149 0.20829674
#> [25,] 0.5926364 0.6100184 0.55966952 0.50394298 0.39373479 0.21306886
#>                V4        V40        V41        V42        V43        V44
#>  [1,] 0.042772306 0.03868274 0.14861552 0.04268471 0.06618703 0.14149278
#>  [2,] 0.057193451 0.29951940 0.67544514 0.76097946 0.46872893 0.44642738
#>  [3,] 0.049897618 0.22997734 0.37957271 0.47323521 0.40940411 0.18579909
#>  [4,] 0.048963026 0.32168077 0.20426003 0.12947743 0.26703308 0.25663523
#>  [5,] 0.106230042 0.07272071 0.06955801 0.27033629 0.23217360 0.01747554
#>  [6,] 0.035693322 0.22552772 0.08002722 0.19868220 0.26988380 0.28754448
#>  [7,] 0.052116065 0.10673563 0.07140918 0.09468404 0.19455070 0.22611198
#>  [8,] 0.102424286 0.31995128 0.33935580 0.31768311 0.17318644 0.09502331
#>  [9,] 0.066476791 0.28397147 0.37563286 0.39467386 0.37237226 0.22556538
#> [10,] 0.058408329 0.39588227 0.41122025 0.38248447 0.28881799 0.18609676
#> [11,] 0.056288642 0.30427545 0.26874107 0.26781543 0.22972017 0.19678385
#> [12,] 0.143291823 0.26795397 0.14993994 0.31924550 0.39906925 0.23408661
#> [13,] 0.086563034 0.21103485 0.54416228 0.64509566 0.48324308 0.46169609
#> [14,] 0.049258058 0.40708137 0.36983338 0.33272257 0.23387860 0.17778073
#> [15,] 0.018813082 0.58937143 0.49729901 0.44167578 0.30311155 0.15139295
#> [16,] 0.102866911 0.36332764 0.32614175 0.26609543 0.07330431 0.19576319
#> [17,] 0.022330778 0.33622911 0.20245837 0.14109806 0.15502831 0.11912085
#> [18,] 0.034950996 0.16630129 0.20468971 0.21189027 0.23466973 0.34684123
#> [19,] 0.033388561 0.32891474 0.33092614 0.25768384 0.16652935 0.18250813
#> [20,] 0.033353310 0.18668136 0.19022950 0.16523808 0.14392182 0.10013978
#> [21,] 0.016941616 0.21516192 0.27991427 0.27229457 0.20532203 0.12045657
#> [22,] 0.006196168 0.78022619 0.67322635 0.61088030 0.51297990 0.38550785
#> [23,] 0.041265571 0.15840093 0.11628539 0.04541427 0.06772659 0.12389620
#> [24,] 0.052256473 0.38990364 0.45454935 0.38081029 0.27562378 0.21082864
#> [25,] 0.037504069 0.16628886 0.15323171 0.05980074 0.05829180 0.03680759
#>              V45        V46        V47        V48         V49         V5
#>  [1,] 0.10560365 0.06555098 0.10580582 0.11217796 0.061600973 0.04352293
#>  [2,] 0.62431359 0.58581906 0.30485822 0.14078950 0.043451521 0.08355989
#>  [3,] 0.20454737 0.20881926 0.19432467 0.16154408 0.099807397 0.06695346
#>  [4,] 0.13171424 0.11268142 0.13463025 0.11381437 0.089805922 0.05976235
#>  [5,] 0.17431196 0.12763182 0.04401636 0.05731113 0.042031672 0.19688105
#>  [6,] 0.29156121 0.22388695 0.14948706 0.10067662 0.058784010 0.06223613
#>  [7,] 0.19611123 0.21580814 0.17607219 0.10781107 0.052543751 0.05860826
#>  [8,] 0.16671138 0.20185796 0.20349263 0.13773343 0.056152590 0.13948629
#>  [9,] 0.10492669 0.08316594 0.07016815 0.04972512 0.021872597 0.17403646
#> [10,] 0.23379941 0.21533068 0.16850488 0.14357073 0.090139383 0.06866362
#> [11,] 0.14141450 0.09648783 0.08888680 0.06625626 0.047934305 0.11943262
#> [12,] 0.07993379 0.07348509 0.03039815 0.03374988 0.048494360 0.13825840
#> [13,] 0.56019734 0.56455982 0.38503068 0.19201919 0.098505325 0.07222492
#> [14,] 0.06723929 0.08610911 0.07912421 0.05955620 0.048291875 0.05081072
#> [15,] 0.10929246 0.04251944 0.07964191 0.07467389 0.040004840 0.03291653
#> [16,] 0.13589753 0.14810856 0.10241226 0.06077764 0.045971008 0.17568699
#> [17,] 0.07408325 0.06388208 0.04113229 0.03415363 0.009037185 0.04798771
#> [18,] 0.32563613 0.21835851 0.11915089 0.04148854 0.023999361 0.04055618
#> [19,] 0.16293312 0.07092259 0.09949415 0.09338464 0.050919211 0.07740642
#> [20,] 0.07752675 0.05821988 0.03814120 0.03525034 0.013874069 0.05935634
#> [21,] 0.09859192 0.10154695 0.05017429 0.03659509 0.030012157 0.03369842
#> [22,] 0.24779383 0.14674390 0.13280465 0.09445516 0.060635411 0.06033177
#> [23,] 0.09796416 0.05272085 0.07086217 0.05204017 0.032038034 0.06009484
#> [24,] 0.23163283 0.22455406 0.20868554 0.17882702 0.117374757 0.06143590
#> [25,] 0.02748005 0.02311689 0.01872348 0.02228985 0.009168117 0.05625548
#>               V50         V51         V52         V53         V54          V55
#>  [1,] 0.022924854 0.012178461 0.013824685 0.001718521 0.006134819 0.0055282934
#>  [2,] 0.034352186 0.042368046 0.039923757 0.023911441 0.018374362 0.0169687798
#>  [3,] 0.023076443 0.016674775 0.010552279 0.003999633 0.003278101 0.0041645132
#>  [4,] 0.027341614 0.034088435 0.008400602 0.014359662 0.013195674 0.0285360167
#>  [5,] 0.042757530 0.027549683 0.026707369 0.024026962 0.017470929 0.0211379750
#>  [6,] 0.015446988 0.024661208 0.013713830 0.006949111 0.006573302 0.0038667726
#>  [7,] 0.012854777 0.016889949 0.012994781 0.004213968 0.004603738 0.0045156322
#>  [8,] 0.017461943 0.019944671 0.019385892 0.011296518 0.007725244 0.0074739678
#>  [9,] 0.012506097 0.012660832 0.012461404 0.012031641 0.009348725 0.0009281183
#> [10,] 0.031998573 0.018429158 0.015521990 0.009400336 0.007249303 0.0134658225
#> [11,] 0.019280824 0.015996092 0.014087939 0.015084152 0.013756756 0.0074539220
#> [12,] 0.026393196 0.021259724 0.009888659 0.015641533 0.039102484 0.0049320757
#> [13,] 0.036694221 0.048092804 0.048921133 0.025723649 0.018605235 0.0123853512
#> [14,] 0.018969523 0.012329079 0.014167529 0.007800905 0.013557702 0.0149433787
#> [15,] 0.013723851 0.007255323 0.003707925 0.019858369 0.015460427 0.0084038961
#> [16,] 0.008178436 0.013806913 0.006358125 0.014710445 0.005166472 0.0154494780
#> [17,] 0.017280805 0.014065724 0.009042210 0.007825232 0.011914486 0.0097165046
#> [18,] 0.017373304 0.013530818 0.002882374 0.006364156 0.011280256 0.0026642312
#> [19,] 0.011192324 0.017434532 0.012614925 0.008264404 0.008367735 0.0065024395
#> [20,] 0.006297747 0.007859465 0.009991560 0.009004016 0.002957646 0.0064438408
#> [21,] 0.009939835 0.005624402 0.009134296 0.003762952 0.003909017 0.0057432049
#> [22,] 0.018661307 0.008555666 0.008803382 0.015790681 0.010430380 0.0088348314
#> [23,] 0.016900164 0.014684885 0.008918985 0.009913103 0.005130177 0.0055239021
#> [24,] 0.032526186 0.014302354 0.016648656 0.014444616 0.009380255 0.0042377379
#> [25,] 0.009969608 0.006500613 0.006431033 0.006086466 0.003665861 0.0026160668
#>               V56           V57         V58          V59         V6         V60
#>  [1,] 0.007866063  1.175924e-02 0.012167537 0.0070888886 0.12240289 0.003984481
#>  [2,] 0.013364154  4.494028e-03 0.010776089 0.0132319946 0.21099809 0.010218549
#>  [3,] 0.008180278  1.592875e-04 0.008388506 0.0098372947 0.07308534 0.003903120
#>  [4,] 0.010096599  1.108270e-02 0.024061212 0.0117057572 0.15407094 0.006683719
#>  [5,] 0.006829969  4.765087e-03 0.015461636 0.0153419319 0.19788457 0.011251193
#>  [6,] 0.003266891  3.307076e-03 0.002618163 0.0039881393 0.08785954 0.007385311
#>  [7,] 0.004101457  3.733561e-03 0.002534796 0.0008287738 0.09964109 0.002985748
#>  [8,] 0.002914147  5.701721e-03 0.005531866 0.0027224417 0.13279108 0.006958877
#>  [9,] 0.001451960  7.580976e-05 0.003534472 0.0074050507 0.09173641 0.002021713
#> [10,] 0.008948702  7.842099e-03 0.011532201 0.0074530933 0.08252798 0.002968895
#> [11,] 0.009438310  7.858240e-03 0.008663076 0.0079322778 0.13548130 0.005750795
#> [12,] 0.007988137 -5.923713e-04 0.007643521 0.0198749920 0.11373089 0.009600889
#> [13,] 0.006639099  1.850996e-02 0.014031214 0.0138165309 0.07666044 0.004037705
#> [14,] 0.004429492  1.216058e-02 0.011815523 0.0083846576 0.04537441 0.003957200
#> [15,] 0.011191873  8.851575e-03 0.008285240 0.0100734442 0.05740510 0.009715859
#> [16,] 0.007408989  1.284561e-02 0.005112453 0.0161154456 0.26492706 0.005186570
#> [17,] 0.010898236  1.243361e-02 0.006728907 0.0076664662 0.07644208 0.008439960
#> [18,] 0.003338441  6.904603e-03 0.006864441 0.0073616340 0.03659971 0.001130629
#> [19,] 0.007387883  9.623892e-03 0.010567554 0.0160320186 0.12784988 0.012636587
#> [20,] 0.006056868  5.742179e-03 0.007911751 0.0075125238 0.10516942 0.007253159
#> [21,] 0.002889151  2.906931e-03 0.003383426 0.0014151970 0.05623322 0.002715854
#> [22,] 0.009722988  6.020175e-03 0.010226202 0.0080834314 0.10344237 0.001682567
#> [23,] 0.007807942  4.981454e-03 0.002389642 0.0040797207 0.09370683 0.003428469
#> [24,] 0.006861387  2.023882e-02 0.010357851 0.0041164354 0.07542262 0.004789531
#> [25,] 0.003928561  4.200126e-03 0.002794070 0.0021099901 0.08505261 0.002194063
#>               V7          V8         V9
#>  [1,] 0.12590831  0.12051519 0.13514099
#>  [2,] 0.15179887  0.16810921 0.25402954
#>  [3,] 0.11324694  0.14123274 0.20408323
#>  [4,] 0.26158350  0.29523385 0.29912075
#>  [5,] 0.17030853  0.10383426 0.13474552
#>  [6,] 0.09311426  0.14734642 0.22571053
#>  [7,] 0.15105500  0.15620930 0.19266369
#>  [8,] 0.12617569  0.08721950 0.14000700
#>  [9,] 0.05277897  0.10059194 0.31937449
#> [10,] 0.10245601  0.14895952 0.24106562
#> [11,] 0.12866770  0.13473134 0.16773166
#> [12,] 0.16862588  0.15924067 0.03787821
#> [13,] 0.03947621  0.10838326 0.15619169
#> [14,] 0.04603996 -0.01010743 0.02065721
#> [15,] 0.03447055  0.05881322 0.04384358
#> [16,] 0.23030439  0.18606408 0.17362761
#> [17,] 0.10897225  0.08817194 0.05539541
#> [18,] 0.09100747  0.08137228 0.07923416
#> [19,] 0.18780124  0.20392071 0.24425595
#> [20,] 0.09974131  0.10295713 0.10515511
#> [21,] 0.07235480  0.07038014 0.10950493
#> [22,] 0.07968019  0.09482233 0.01620581
#> [23,] 0.14305310  0.10210643 0.09853195
#> [24,] 0.17677588  0.26206162 0.25561629
#> [25,] 0.08723743  0.09021096 0.15790339
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551