Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.017698340 0.21456836 0.28263878 0.31461057 0.28026816 0.20426685
#>  [2,] 0.045918856 0.05767241 0.14823396 0.21133159 0.26591930 0.30604942
#>  [3,] 0.035788428 0.24794401 0.30389045 0.35728429 0.33483239 0.26379830
#>  [4,] 0.018712435 0.11132150 0.24370592 0.27631758 0.25884209 0.16018973
#>  [5,] 0.103110492 0.57839249 0.47059305 0.20169006 0.20024012 0.28639491
#>  [6,] 0.043343063 0.25523058 0.25935552 0.30011507 0.35720309 0.37417040
#>  [7,] 0.073100000 0.11140000 0.17390000 0.31600000 0.32490000 0.21640000
#>  [8,] 0.036899419 0.68286788 0.70775605 0.49051551 0.31684344 0.21729056
#>  [9,] 0.043873373 0.40964074 0.55178734 0.58389600 0.65092448 0.67167079
#> [10,] 0.038757247 0.16293230 0.23278635 0.33120226 0.36053458 0.41534667
#> [11,] 0.029690662 0.23955216 0.29833823 0.32308841 0.26450223 0.17613599
#> [12,] 0.025258341 0.36124136 0.44970374 0.48424759 0.45910300 0.37791480
#> [13,] 0.035168727 0.18062137 0.17380603 0.19649169 0.28724159 0.28497816
#> [14,] 0.019260056 0.10968474 0.14731679 0.12416760 0.13742412 0.11963017
#> [15,] 0.004277210 0.19742494 0.17449371 0.23141757 0.40133765 0.29462007
#> [16,] 0.009827953 0.10566066 0.10243962 0.11071173 0.18698358 0.17100660
#> [17,] 0.051917795 0.31388105 0.27992228 0.10434628 0.20082804 0.48874741
#> [18,] 0.020953731 0.09673570 0.12202726 0.15294105 0.13720944 0.15459938
#> [19,] 0.031700000 0.35130000 0.17860000 0.06580000 0.05130000 0.37520000
#> [20,] 0.034639911 0.05727519 0.08566820 0.10992152 0.18933726 0.24210269
#> [21,] 0.021274771 0.08495024 0.05937394 0.07036703 0.10850071 0.03077067
#> [22,] 0.011075761 0.06321696 0.06346360 0.05169548 0.09756369 0.11253367
#> [23,] 0.013658737 0.17550964 0.13458337 0.14618397 0.24001303 0.33778233
#> [24,] 0.020789781 0.14852775 0.13654957 0.23835009 0.13863756 0.20420746
#> [25,] 0.025300000 0.21970000 0.26530000 0.32230000 0.55820000 0.69160000
#>               V15          V16        V17        V18       V19           V2
#>  [1,]  0.08067522  0.246741291 0.42590038 0.61957871 0.7736143  0.006638179
#>  [2,]  0.35943097  0.363044337 0.51323583 0.64643961 0.8370695  0.069478705
#>  [3,]  0.18848892  0.203597968 0.26079447 0.40493173 0.5725480  0.046577318
#>  [4,]  0.07602205  0.151133320 0.26338558 0.35959739 0.4354800  0.035307907
#>  [5,]  0.55243411  0.570092626 0.22263003 0.16661841 0.3941755  0.098584890
#>  [6,]  0.33398428  0.425347085 0.41937567 0.30175540 0.4402772  0.074169050
#>  [7,]  0.20310000  0.258000000 0.17960000 0.24220000 0.3609000  0.124900000
#>  [8,]  0.24918662  0.293128577 0.23315171 0.29656373 0.4181081  0.071582290
#>  [9,]  0.69684450  0.762387935 0.90961555 0.93574854 0.8904695  0.035768841
#> [10,]  0.53485471  0.602021683 0.68172454 0.78953298 0.8836350  0.048344507
#> [11,]  0.17871939  0.192997995 0.19820466 0.23980409 0.2666091  0.034329479
#> [12,]  0.31228532  0.350018016 0.38659225 0.44807437 0.5803256  0.021750835
#> [13,]  0.25482418  0.271656230 0.28932740 0.17436412 0.2020222  0.056747398
#> [14,]  0.06256467  0.086474642 0.11940512 0.18271022 0.3997650  0.013075246
#> [15,]  0.28500219  0.527460090 0.40795472 0.15821289 0.3102454 -0.006284139
#> [16,]  0.17285930  0.218191702 0.24004440 0.30157213 0.4257517  0.013645814
#> [17,]  0.65452364  0.760926761 0.82742465 0.72226841 0.5080054  0.061604248
#> [18,]  0.14587268  0.221851420 0.31375632 0.33753155 0.3053050  0.022205648
#> [19,]  0.54190000  0.544000000 0.51500000 0.42620000 0.2024000  0.095600000
#> [20,]  0.19376080  0.132322227 0.16313774 0.07739285 0.1128801  0.065091412
#> [21,] -0.01757867 -0.005405404 0.04843102 0.21643105 0.3667880  0.001328145
#> [22,]  0.17292978  0.264497444 0.32677840 0.38212143 0.4210491  0.013654165
#> [23,]  0.37473831  0.256523890 0.08115961 0.16825027 0.2500403  0.023696927
#> [24,]  0.20211289  0.255550810 0.20201445 0.32363668 0.4130954  0.030477943
#> [25,]  0.79430000  0.715200000 0.35120000 0.20080000 0.2676000  0.080800000
#>             V20       V21       V22       V23        V24       V25       V26
#>  [1,] 0.7736507 0.8024096 0.8481455 0.9331268 0.92862082 0.9392008 0.7809281
#>  [2,] 0.9149055 0.8833648 0.9897504 1.0012447 0.93681648 0.8407473 0.7071191
#>  [3,] 0.6788722 0.7868093 0.7808721 0.8355959 0.94956926 0.9651791 0.9901265
#>  [4,] 0.5637048 0.6669903 0.8026318 0.8616781 0.79572741 0.8385361 0.9440995
#>  [5,] 0.8633907 0.8880686 0.3910093 0.1320484 0.15683210 0.2084159 0.3182278
#>  [6,] 0.5589242 0.6143125 0.5862059 0.5549668 0.62464489 0.6024387 0.6962547
#>  [7,] 0.1810000 0.2604000 0.6572000 0.9734000 0.97570000 0.8079000 0.6521000
#>  [8,] 0.5655083 0.7018692 0.8517922 0.9462408 0.98058022 0.7166657 0.5994064
#>  [9,] 0.8358102 0.7160852 0.4380578 0.1231232 0.06429345 0.2549836 0.3484626
#> [10,] 0.9179425 0.8521141 0.8418231 0.8975857 0.76857152 0.4391097 0.2615339
#> [11,] 0.2825121 0.3965176 0.5293063 0.6611747 0.76979805 0.8308865 0.8778850
#> [12,] 0.7016022 0.7923794 0.8715853 0.8745417 0.88388971 0.8665451 0.8597491
#> [13,] 0.2675696 0.3118467 0.4540712 0.5385890 0.54344897 0.5353023 0.5600561
#> [14,] 0.3806321 0.2392760 0.2627555 0.4167382 0.58271480 0.5549204 0.5278690
#> [15,] 0.3813642 0.3510210 0.4451931 0.6319113 0.45244628 0.6201471 0.9069366
#> [16,] 0.5373509 0.5676168 0.5501190 0.5975660 0.65201275 0.7526044 0.8387591
#> [17,] 0.3809627 0.5420905 0.7068387 0.8084120 0.89587437 0.8663469 0.6836661
#> [18,] 0.3930908 0.4577618 0.5322886 0.6019098 0.66897172 0.6414169 0.6420447
#> [19,] 0.4233000 0.7723000 0.9735000 0.9390000 0.55590000 0.5268000 0.6826000
#> [20,] 0.1457170 0.1559812 0.2539876 0.1328410 0.17632767 0.4136364 0.6269511
#> [21,] 0.4136059 0.4096308 0.4947621 0.6438362 0.76177910 0.8768427 0.8771045
#> [22,] 0.4165656 0.4432359 0.4091947 0.3308847 0.40092995 0.4402170 0.5068639
#> [23,] 0.3087371 0.2854787 0.2867866 0.5602490 0.75229061 0.7243986 0.7067974
#> [24,] 0.3545298 0.2603681 0.2916485 0.4190174 0.36889026 0.2894726 0.4338438
#> [25,] 0.4299000 0.5280000 0.3489000 0.1430000 0.54530000 0.6338000 0.7712000
#>             V27       V28       V29         V3       V30       V31        V32
#>  [1,] 0.4669456 0.3921680 0.3803621 0.01551168 0.5600954 0.7228762 0.82343697
#>  [2,] 0.6616354 0.5274420 0.3701250 0.16589841 0.2881893 0.1440161 0.04450687
#>  [3,] 0.9954917 1.0663295 0.9869704 0.04145428 0.7218205 0.4830934 0.28744917
#>  [4,] 1.0015224 0.9239091 0.7452667 0.03581120 0.5687921 0.2394901 0.29276245
#>  [5,] 0.2273892 0.5022306 0.5379398 0.04242377 0.6708349 0.3459866 0.39038663
#>  [6,] 0.8420910 0.8720273 0.7899593 0.06805056 0.7814730 0.7052254 0.68853400
#>  [7,] 0.4915000 0.5363000 0.7649000 0.16650000 0.5250000 0.5101000 0.42190000
#>  [8,] 0.5159048 0.4651235 0.4616701 0.07297033 0.4077974 0.4172310 0.20070885
#>  [9,] 0.3857215 0.3549121 0.1953973 0.08184541 0.1252512 0.2067276 0.22845172
#> [10,] 0.2706998 0.2985754 0.2823772 0.05523012 0.2659572 0.2272917 0.33660616
#> [11,] 0.9148603 0.9416439 0.8759918 0.03197636 0.7443851 0.6882269 0.59343227
#> [12,] 0.9091427 0.9527320 0.8125533 0.02730587 0.5523923 0.3600283 0.34209266
#> [13,] 0.6250881 0.7954076 0.8552253 0.08188196 0.8484664 0.8634260 0.81572623
#> [14,] 0.4505235 0.6743321 0.9839155 0.01659067 1.0041860 0.9568846 0.88670998
#> [15,] 0.8496107 0.7643200 0.5056415 0.01814637 0.4514289 0.5411512 0.54643462
#> [16,] 0.8468069 0.7909626 0.6248292 0.01677524 0.4621959 0.4031285 0.32748304
#> [17,] 0.6635235 0.6031798 0.4214214 0.06469545 0.4807117 0.4826364 0.31591179
#> [18,] 0.6767611 0.7390925 0.8105522 0.02356844 0.7624950 0.6643704 0.57300032
#> [19,] 0.5713000 0.5429000 0.2177000 0.13210000 0.2149000 0.5811000 0.63230000
#> [20,] 0.6292655 0.5875415 0.5546109 0.03550859 0.5988807 0.7274623 0.79006763
#> [21,] 0.7723934 0.8221175 0.9688780 0.01329942 0.9665608 0.9280094 0.90021996
#> [22,] 0.7439425 0.9503962 1.0055925 0.01191172 0.9121602 0.8350880 0.76420195
#> [23,] 0.6958666 0.8421511 0.9912192 0.03711646 0.8718841 0.7016122 0.58357678
#> [24,] 0.4060727 0.3439803 0.3935447 0.06066237 0.4801173 0.6314175 0.72242403
#> [25,] 0.6838000 0.8015000 0.8073000 0.05070000 0.8310000 0.7792000 0.50490000
#>              V33       V34        V35        V36       V37        V38
#>  [1,] 0.91583974 0.8671736 0.65476393 0.43532942 0.1324207 0.19134407
#>  [2,] 0.09871832 0.2087822 0.44338233 0.52210322 0.4344239 0.43764738
#>  [3,] 0.20786254 0.1513174 0.17665952 0.08555205 0.0352198 0.12698260
#>  [4,] 0.34831227 0.5158281 0.76405859 0.92960535 0.8336388 0.62293993
#>  [5,] 0.66070964 0.7827083 0.88983337 0.84944254 0.8476192 0.88148164
#>  [6,] 0.45793637 0.3404953 0.26005818 0.21530361 0.2872502 0.38079362
#>  [7,] 0.41600000 0.1906000 0.02230000 0.42190000 0.5496000 0.24830000
#>  [8,] 0.48607472 0.6637582 0.60490162 0.45710866 0.4346954 0.37003045
#>  [9,] 0.28098871 0.2464262 0.33035249 0.05365465 0.1490932 0.16282546
#> [10,] 0.32052700 0.2596193 0.22921553 0.26612412 0.3563626 0.48617970
#> [11,] 0.46236423 0.3146460 0.18948331 0.12031349 0.1667253 0.16378515
#> [12,] 0.29218303 0.1594316 0.09302589 0.17549698 0.2233967 0.30640412
#> [13,] 0.71511118 0.5727402 0.54601020 0.41638144 0.2671120 0.34164893
#> [14,] 0.76886265 0.6550778 0.53903441 0.44738463 0.4586977 0.18154096
#> [15,] 0.39073393 0.3545829 0.75094799 0.78886209 0.3934574 0.06247006
#> [16,] 0.28392660 0.2952087 0.35591391 0.37793417 0.3539779 0.28174350
#> [17,] 0.25781307 0.2266890 0.53978553 0.67801876 0.6390277 0.42256594
#> [18,] 0.60587867 0.6731036 0.65349312 0.64016410 0.5977466 0.48246095
#> [19,] 0.29650000 0.1873000 0.29690000 0.51630000 0.6153000 0.42830000
#> [20,] 0.78982151 0.7320312 0.88848804 1.02383829 0.9279153 0.77022098
#> [21,] 0.99927963 0.9975902 0.80520867 0.51407536 0.2338071 0.16350176
#> [22,] 0.72816625 0.6146169 0.56881706 0.58099747 0.5360001 0.36536175
#> [23,] 0.65593471 0.5602836 0.37652973 0.28796679 0.1815708 0.17347757
#> [24,] 0.64299346 0.4650008 0.49402891 0.48125189 0.6674414 0.91201409
#> [25,] 0.14130000 0.2767000 0.50840000 0.47870000 0.1356000 0.22990000
#>              V39           V4        V40        V41       V42       V43
#>  [1,] 0.38329424 0.0427605276 0.54111572 0.51233861 0.3615406 0.2055105
#>  [2,] 0.41045984 0.2410836280 0.33341284 0.41273392 0.5613447 0.4198016
#>  [3,] 0.15165440 0.0241594135 0.18431217 0.09298503 0.1299040 0.2515665
#>  [4,] 0.44987494 0.0522466591 0.31340172 0.35156798 0.3548236 0.4292737
#>  [5,] 0.74476636 0.0834321523 0.52171329 0.26367161 0.2125978 0.2036841
#>  [6,] 0.40790500 0.0555176001 0.26394811 0.48119666 0.5996057 0.5184559
#>  [7,] 0.20340000 0.1496000000 0.27290000 0.28370000 0.4463000 0.3178000
#>  [8,] 0.43751286 0.0769250811 0.42141209 0.28767447 0.3650658 0.3600844
#>  [9,] 0.14826015 0.1342753776 0.08389663 0.04561943 0.1340243 0.1642014
#> [10,] 0.41122487 0.0591862491 0.29042756 0.25726039 0.2639331 0.2834168
#> [11,] 0.20222005 0.0418919328 0.20473803 0.26785009 0.2216034 0.2053305
#> [12,] 0.40936795 0.0624652442 0.36425459 0.27007333 0.2961940 0.3299799
#> [13,] 0.30827828 0.0830701894 0.29693993 0.55293695 0.5671070 0.3843307
#> [14,] 0.05103457 0.0399498353 0.13026848 0.15365479 0.1510955 0.2496195
#> [15,] 0.36448027 0.0418597091 0.62221683 0.34673140 0.2053063 0.1966861
#> [16,] 0.23593996 0.0254815160 0.19967506 0.21530838 0.1690559 0.1312384
#> [17,] 0.20303607 0.0509268967 0.32130794 0.33420573 0.3227722 0.2637320
#> [18,] 0.36165388 0.0298262609 0.34242619 0.31108455 0.2299957 0.1749588
#> [19,] 0.54790000 0.1408000000 0.61330000 0.50170000 0.2377000 0.1957000
#> [20,] 0.57583401 0.0281059054 0.40128817 0.31197835 0.2816216 0.2920458
#> [21,] 0.10950388 0.0004974247 0.29825714 0.37125016 0.4082322 0.3251852
#> [22,] 0.15942642 0.0047428101 0.07558258 0.04961456 0.1797808 0.2635516
#> [23,] 0.26054044 0.0296279751 0.17225230 0.15206192 0.2180703 0.2339587
#> [24,] 0.96187929 0.0253568233 0.87868417 0.67692113 0.5700035 0.3910179
#> [25,] 0.27890000 0.0244000000 0.38330000 0.29330000 0.1155000 0.1705000
#>              V44        V45        V46        V47        V48        V49
#>  [1,] 0.09295308 0.08068088 0.11109578 0.12099401 0.11035230 0.08339636
#>  [2,] 0.17299125 0.19051660 0.14926691 0.08625115 0.08617251 0.04836751
#>  [3,] 0.24647463 0.24098090 0.22715678 0.16387376 0.09939116 0.06447639
#>  [4,] 0.37398084 0.38185548 0.27979417 0.11341051 0.02685520 0.04167034
#>  [5,] 0.28308885 0.44965278 0.33379637 0.15249139 0.10124114 0.06359024
#>  [6,] 0.43308682 0.51043126 0.45044300 0.30149085 0.22025677 0.13555588
#>  [7,] 0.08070000 0.11920000 0.21340000 0.32410000 0.29450000 0.14740000
#>  [8,] 0.19316716 0.15088746 0.03987938 0.10957027 0.17598526 0.08954616
#>  [9,] 0.07317102 0.10363857 0.05854151 0.04751224 0.06537584 0.03806801
#> [10,] 0.17784890 0.05718343 0.02773037 0.01725749 0.01270514 0.01971188
#> [11,] 0.14889557 0.11973922 0.10345566 0.12786452 0.11719400 0.06545096
#> [12,] 0.31831411 0.29091345 0.16826725 0.11232800 0.09574624 0.06070512
#> [13,] 0.46486167 0.55986514 0.49653266 0.31999370 0.16275713 0.07780588
#> [14,] 0.37822476 0.37217438 0.23617172 0.09995558 0.01358716 0.01101599
#> [15,] 0.30634233 0.16222777 0.12756376 0.03318055 0.04550546 0.00440597
#> [16,] 0.09114758 0.07685005 0.07303855 0.05144811 0.03932558 0.02243024
#> [17,] 0.22204722 0.23076547 0.19048564 0.16881988 0.15000936 0.09286183
#> [18,] 0.14614307 0.12797498 0.11490687 0.07677001 0.04449382 0.01893318
#> [19,] 0.17490000 0.13040000 0.05970000 0.11240000 0.10470000 0.05070000
#> [20,] 0.27495788 0.23282435 0.15810354 0.09932523 0.08947368 0.05039850
#> [21,] 0.27960494 0.25512887 0.19011167 0.09152224 0.03107127 0.02679945
#> [22,] 0.28849118 0.23556830 0.16741899 0.10076018 0.12097636 0.07153395
#> [23,] 0.15376440 0.01870716 0.03489434 0.04154530 0.02857616 0.01054520
#> [24,] 0.32201957 0.37780361 0.35204821 0.26982920 0.17428780 0.08953535
#> [25,] 0.12940000 0.09090000 0.08000000 0.05670000 0.01980000 0.01140000
#>                V5         V50         V51         V52         V53         V54
#>  [1,] 0.057116267 0.025221924 0.013870985 0.009127955 0.004926334 0.002597469
#>  [2,] 0.303331046 0.016233763 0.029152692 0.022947832 0.020628376 0.023893512
#>  [3,] 0.110196283 0.020169368 0.022952914 0.015005921 0.012392754 0.001814936
#>  [4,] 0.041044104 0.015123373 0.004016307 0.014168672 0.007364220 0.003735853
#>  [5,] 0.072869561 0.033994754 0.018892469 0.016322744 0.013449730 0.023568356
#>  [6,] 0.111274329 0.053154624 0.046488301 0.023895771 0.013112546 0.017241943
#>  [7,] 0.144300000 0.021100000 0.036100000 0.044400000 0.023000000 0.029000000
#>  [8,] 0.026975500 0.020386102 0.017834777 0.016311829 0.016370201 0.020440983
#>  [9,] 0.113432384 0.026539485 0.031657279 0.026057082 0.027438898 0.012151359
#> [10,] 0.096154300 0.008383824 0.012418168 0.015565121 0.009357780 0.016524859
#> [11,] 0.051610478 0.015708064 0.015998767 0.014191728 0.008713030 0.010504450
#> [12,] 0.113587006 0.018005422 0.016564269 0.015122223 0.013751357 0.012527453
#> [13,] 0.071985670 0.032223820 0.031939922 0.034460479 0.018925535 0.016332009
#> [14,] 0.037016743 0.017967261 0.012260506 0.002367592 0.006753813 0.012397333
#> [15,] 0.117671828 0.002461891 0.018269825 0.003504919 0.008337335 0.019173953
#> [16,] 0.024138284 0.013886458 0.010868731 0.011642476 0.007381529 0.009548020
#> [17,] 0.074753122 0.027212923 0.013368081 0.016046032 0.010052537 0.012574974
#> [18,] 0.048220637 0.016334188 0.007660098 0.008404846 0.010047466 0.009875658
#> [19,] 0.167400000 0.015900000 0.019500000 0.020100000 0.024800000 0.013100000
#> [20,] 0.037923963 0.003521215 0.007435874 0.013931851 0.024187430 0.007081063
#> [21,] 0.005319059 0.010751774 0.006594607 0.010952246 0.006541840 0.006355896
#> [22,] 0.015899134 0.016698712 0.010632026 0.009005094 0.001227519 0.008459522
#> [23,] 0.021570589 0.018095585 0.016031524 0.014165901 0.012866483 0.005372329
#> [24,] 0.033397307 0.037002797 0.024859835 0.013419619 0.008585855 0.012319341
#> [25,] 0.172400000 0.015100000 0.008500000 0.017800000 0.007300000 0.007900000
#>                V55         V56          V57         V58         V59          V6
#>  [1,] 0.0076155965 0.005057109 0.0025290432 0.007340871 0.006623905 0.096792912
#>  [2,] 0.0067539896 0.013498861 0.0161210110 0.014817129 0.015175736 0.175357428
#>  [3,] 0.0074179841 0.003751994 0.0048517470 0.007398749 0.003940664 0.151200322
#>  [4,] 0.0056939468 0.001208850 0.0014032163 0.002726875 0.003851855 0.080958490
#>  [5,] 0.0231198838 0.010625906 0.0113822597 0.013389435 0.009821736 0.104380646
#>  [6,] 0.0165233532 0.015103725 0.0113430901 0.016265358 0.010414747 0.103494877
#>  [7,] 0.0141000000 0.016100000 0.0177000000 0.019400000 0.020700000 0.277000000
#>  [8,] 0.0133708468 0.006771493 0.0040124757 0.016029864 0.008354816 0.040606011
#>  [9,] 0.0191827629 0.017184273 0.0072413163 0.030282588 0.014614872 0.164496741
#> [10,] 0.0181511271 0.015533465 0.0081670553 0.009933498 0.010009688 0.083055478
#> [11,] 0.0063337712 0.007517001 0.0046707842 0.005294513 0.007248935 0.104041178
#> [12,] 0.0065781026 0.009743922 0.0054369502 0.006953274 0.006906053 0.152538808
#> [13,] 0.0141378180 0.009816914 0.0123994952 0.009611343 0.013459238 0.103567790
#> [14,] 0.0009019761 0.001854787 0.0058481270 0.004167639 0.006806473 0.020117959
#> [15,] 0.0055195260 0.014874586 0.0150583855 0.003500491 0.018066052 0.154485798
#> [16,] 0.0090662341 0.006820465 0.0049581247 0.007216080 0.003254694 0.052829111
#> [17,] 0.0079888677 0.004975883 0.0154731965 0.008792200 0.008520144 0.118146742
#> [18,] 0.0103034752 0.006181022 0.0084295621 0.004305655 0.007367204 0.090042309
#> [19,] 0.0070000000 0.013800000 0.0092000000 0.014300000 0.003600000 0.171000000
#> [20,] 0.0178533376 0.009623467 0.0056286790 0.011187571 0.012417556 0.080170861
#> [21,] 0.0028867374 0.000448372 0.0001895582 0.002125987 0.006190999 0.002813438
#> [22,] 0.0065432022 0.002519580 0.0029060342 0.003619248 0.007025468 0.106021091
#> [23,] 0.0078844164 0.006819554 0.0058734664 0.003927444 0.012266229 0.072328027
#> [24,] 0.0096678455 0.007587814 0.0041236214 0.008750017 0.006510692 0.057289555
#> [25,] 0.0038000000 0.011600000 0.0033000000 0.003900000 0.008100000 0.382300000
#>               V60         V7          V8          V9
#>  [1,] 0.003526489 0.09872355 0.171659495 0.233303141
#>  [2,] 0.008409232 0.06340329 0.199092574 0.063227814
#>  [3,] 0.003938141 0.14470151 0.183918198 0.219450063
#>  [4,] 0.005450051 0.06863881 0.055770643 0.042124064
#>  [5,] 0.010969973 0.23985163 0.401307246 0.543036331
#>  [6,] 0.007998861 0.13180322 0.160122007 0.242216588
#>  [7,] 0.005700000 0.25550000 0.171200000 0.046600000
#>  [8,] 0.003655527 0.09386806 0.239921588 0.422459791
#>  [9,] 0.015314824 0.19089807 0.143226312 0.266275020
#> [10,] 0.008534306 0.08589503 0.113936501 0.152258870
#> [11,] 0.007369020 0.12296364 0.125889744 0.159613522
#> [12,] 0.004736716 0.14373299 0.182906486 0.265515138
#> [13,] 0.006644831 0.07525529 0.109540877 0.215841513
#> [14,] 0.001014833 0.08706304 0.093918676 0.074456881
#> [15,] 0.009425922 0.13734914 0.074872726 0.089293471
#> [16,] 0.003068812 0.07380013 0.058923504 0.104518993
#> [17,] 0.006474868 0.20998420 0.252787170 0.282226069
#> [18,] 0.005321481 0.09204414 0.094078399 0.097028653
#> [19,] 0.010300000 0.07310000 0.140100000 0.208300000
#> [20,] 0.005832763 0.09758706 0.059647578 0.008032636
#> [21,] 0.004513335 0.04259936 0.007384361 0.024315199
#> [22,] 0.004143992 0.10317725 0.040037875 0.083295584
#> [23,] 0.007730945 0.09524547 0.122247675 0.157963356
#> [24,] 0.011277935 0.12631265 0.144772012 0.084012953
#> [25,] 0.005300000 0.37290000 0.358300000 0.342900000
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841