Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10         V11        V12        V13        V14
#>  [1,] 0.031007685  0.27845807  0.30830487 0.33756878 0.34291022 0.30837580
#>  [2,] 0.026064122  0.29791813  0.33131864 0.33195105 0.31115634 0.28529793
#>  [3,] 0.066509411  0.13622057  0.22272092 0.18928429 0.24057402 0.30850266
#>  [4,] 0.056997866  0.20869925  0.31404929 0.42485369 0.50523500 0.44730509
#>  [5,] 0.026300245  0.37759607  0.37884697 0.11358486 0.46741038 1.05523900
#>  [6,] 0.061849176  0.13092215  0.26181953 0.27481658 0.40999673 0.35309470
#>  [7,] 0.062019729  0.39976124  0.59276127 0.53270629 0.52435501 0.51299672
#>  [8,] 0.118633048  0.20086449  0.28322741 0.22995525 0.29803970 0.33483620
#>  [9,] 0.012224209  0.36711332  0.43608112 0.36465559 0.31903140 0.18989028
#> [10,] 0.023259692  0.20112370  0.16946166 0.10626027 0.12000035 0.14867878
#> [11,] 0.039730007  0.24903985  0.27369790 0.28023513 0.29969817 0.31670102
#> [12,] 0.095431998  0.54290838  0.39958967 0.11847145 0.13452649 0.23284061
#> [13,] 0.057700441  0.06786090  0.19161744 0.35187512 0.48463401 0.53612325
#> [14,] 0.026200000  0.61940000  0.63330000 0.70600000 0.55440000 0.53200000
#> [15,] 0.029100000  0.11460000  0.09420000 0.07940000 0.02520000 0.11910000
#> [16,] 0.004261281 -0.08083977 -0.01579217 0.04141952 0.13017873 0.15119872
#> [17,] 0.030766110  0.18473795  0.23509509 0.21147366 0.31883279 0.44204164
#> [18,] 0.015115768  0.08821647  0.08435369 0.05848465 0.08490859 0.09950179
#> [19,] 0.009150472  0.14918055  0.11912476 0.16792423 0.24432865 0.13091312
#> [20,] 0.024055475  0.20851852  0.19490044 0.13366696 0.17572840 0.22369457
#> [21,] 0.008268492  0.10057384  0.09104191 0.05511746 0.07687985 0.08835694
#> [22,] 0.011194578  0.09420164  0.04802291 0.01086957 0.03241861 0.12055605
#> [23,] 0.012167543  0.11675344  0.18250757 0.32998492 0.31887927 0.24781043
#> [24,] 0.034343372  0.29664791  0.24131081 0.22209767 0.24860849 0.14592384
#> [25,] 0.014718367  0.11128289  0.12790895 0.10372250 0.13194419 0.14500790
#>              V15        V16       V17       V18       V19           V2
#>  [1,] 0.24179194 0.26379546 0.2844686 0.2886157 0.3862487 0.0369787467
#>  [2,] 0.27088104 0.31560782 0.3686190 0.4788560 0.5630132 0.0290566814
#>  [3,] 0.45553877 0.54770870 0.5279886 0.3477944 0.4567516 0.1235379698
#>  [4,] 0.54080441 0.67992049 0.8933594 0.9263842 0.9097231 0.1196575662
#>  [5,] 1.00847154 0.86978135 0.6938324 0.8103549 0.9262097 0.0352225882
#>  [6,] 0.23978817 0.28380782 0.3394880 0.1649045 0.2872082 0.0920863247
#>  [7,] 0.61536027 0.80414433 0.9468002 0.9478017 0.8250043 0.0964461240
#>  [8,] 0.37594447 0.41379166 0.4412200 0.4570555 0.5049390 0.2120745682
#>  [9,] 0.12269297 0.18255072 0.3219610 0.4447727 0.6224868 0.0123032038
#> [10,] 0.35607751 0.50932095 0.5989810 0.5429755 0.6107527 0.0280712104
#> [11,] 0.22401515 0.17384065 0.2426284 0.1891970 0.2488927 0.0446033397
#> [12,] 0.43635055 0.41729582 0.2488704 0.1963376 0.3866840 0.1050540950
#> [13,] 0.55972137 0.56274742 0.6778291 0.7804219 0.8913492 0.0562676014
#> [14,] 0.64790000 0.69310000 0.6759000 0.7551000 0.8929000 0.0582000000
#> [15,] 0.10450000 0.20500000 0.1556000 0.2690000 0.3784000 0.0400000000
#> [16,] 0.03407724 0.13364483 0.3307180 0.3544109 0.3784750 0.0009063757
#> [17,] 0.57046267 0.62881602 0.5550167 0.5111085 0.5488392 0.0483685509
#> [18,] 0.02903376 0.12355412 0.1719809 0.4036447 0.3416531 0.0129002851
#> [19,] 0.19005726 0.22569016 0.2056206 0.2612716 0.2629255 0.0405855607
#> [20,] 0.31291890 0.62531025 0.7478204 0.6355148 0.3730367 0.0179313344
#> [21,] 0.11878607 0.12949380 0.1345186 0.1618713 0.2208498 0.0149412018
#> [22,] 0.07586701 0.09728161 0.2030976 0.2988037 0.3791193 0.0093320578
#> [23,] 0.22506189 0.48605628 0.7202402 0.7210643 0.5354601 0.0281150604
#> [24,] 0.35003058 0.33070756 0.2269357 0.2155423 0.2464959 0.0105954652
#> [25,] 0.15994091 0.17845275 0.2024524 0.2626108 0.3388468 0.0164316034
#>             V20       V21       V22         V23         V24        V25
#>  [1,] 0.4758302 0.5258319 0.5675079  0.62472324  0.67779961 0.70732281
#>  [2,] 0.6122953 0.7176795 0.7956648  0.82682241  0.77824657 0.70600839
#>  [3,] 0.5693655 0.7170801 0.7238010  0.55151077  0.62365442 0.64671456
#>  [4,] 0.8234320 0.8178709 0.5699177  0.32098166  0.36020908 0.24801957
#>  [5,] 0.8428125 0.8585033 0.9558580  1.04245201  0.94601624 0.80781852
#>  [6,] 0.3474919 0.4780483 0.6206697  0.64259636  0.59518515 0.48982637
#>  [7,] 0.7811794 0.7529477 0.3578641 -0.05781436 -0.03838769 0.26994689
#>  [8,] 0.3613115 0.2260315 0.1876631  0.41341237  0.63767387 0.49229403
#>  [9,] 0.6727834 0.7583988 0.8577337  0.87013778  0.91839705 0.98494906
#> [10,] 0.7930186 0.9379980 1.0251455  0.88704873  0.63732155 0.41175938
#> [11,] 0.3701019 0.4357767 0.5171387  0.57333547  0.63266870 0.61156746
#> [12,] 0.8589787 0.9464181 0.4258666  0.10341053  0.20028116 0.13968362
#> [13,] 0.8919439 0.9261337 0.9115488  0.97715150  0.99503302 0.70715910
#> [14,] 0.8619000 0.7974000 0.6737000  0.42930000  0.36480000 0.53310000
#> [15,] 0.4024000 0.3470000 0.1395000  0.12080000  0.28270000 0.15000000
#> [16,] 0.5216838 0.6902296 0.7978338  0.92660824  0.96460182 0.91613485
#> [17,] 0.5422601 0.5238840 0.5152765  0.62379173  0.76292143 0.70655814
#> [18,] 0.2997595 0.2014672 0.1941943  0.18855754  0.21626307 0.03960283
#> [19,] 0.1285159 0.2163127 0.1904735  0.21251991  0.23901606 0.24607292
#> [20,] 0.3504094 0.4546759 0.5462376  0.60478252  0.66324686 0.81272969
#> [21,] 0.2476025 0.2851399 0.3440096  0.50421256  0.70187383 0.89156377
#> [22,] 0.5512374 0.6399065 0.7650366  0.80661254  0.73799785 0.82005630
#> [23,] 0.3748345 0.4095007 0.4632681  0.57986767  0.72754318 0.84699241
#> [24,] 0.4275845 0.3151467 0.6336257  0.59976546  0.32708198 0.23966384
#> [25,] 0.3847716 0.4136950 0.4892406  0.58083986  0.68160807 0.68716693
#>              V26       V27        V28        V29           V3       V30
#>  [1,] 0.80318203 0.8790829 0.86650526 0.74079087  0.049715879 0.6499730
#>  [2,] 0.79973193 0.9067334 0.94122763 0.86784557  0.044020854 0.6636854
#>  [3,] 0.67723140 0.7520975 0.82062690 0.85265768  0.131503196 0.9276499
#>  [4,] 0.09064938 0.1001086 0.13948953 0.16240568  0.126993761 0.1200399
#>  [5,] 0.82185417 0.6089682 0.41920839 0.57144921  0.014450121 0.3795359
#>  [6,] 0.51624820 0.5845038 0.81627211 0.87387463  0.061286707 0.8195409
#>  [7,] 0.29269578 0.2418325 0.22298024 0.01144189  0.156793981 0.2012442
#>  [8,] 0.29424901 0.4298555 0.58549979 0.51624898  0.277365184 0.4187113
#>  [9,] 0.98969718 0.9384087 0.77948041 0.50706885  0.013136434 0.4026575
#> [10,] 0.27014750 0.1712089 0.09455004 0.25042905  0.012163316 0.4292155
#> [11,] 0.57087671 0.6295857 0.76372084 0.85638697  0.085223876 0.9150307
#> [12,] 0.29553590 0.1953907 0.38280478 0.47827273  0.080618094 0.6721124
#> [13,] 0.53751360 0.4469041 0.39699771 0.37225809  0.069794573 0.3427082
#> [14,] 0.24130000 0.5070000 0.85330000 0.60360000  0.109900000 0.8514000
#> [15,] 0.26260000 0.4468000 0.75200000 0.90360000  0.077100000 0.7812000
#> [16,] 0.69639292 0.8079838 0.88060211 0.68806860  0.036369841 0.6269030
#> [17,] 0.73341335 0.7424294 0.73046881 0.64085177  0.057986939 0.6484975
#> [18,] 0.17623095 0.4111171 0.63542494 0.78034020 -0.004899505 0.7184665
#> [19,] 0.30657352 0.1955746 0.47374447 0.64444652  0.030710684 0.6380790
#> [20,] 0.93836602 0.9471702 0.85601231 0.63868953  0.024233291 0.6684764
#> [21,] 0.88963130 0.7872258 0.64612452 0.50600420  0.006180180 0.3535927
#> [22,] 0.93042263 0.9753253 0.86428434 0.73114803  0.010682718 0.6302424
#> [23,] 0.89676149 0.9255967 0.90330797 0.92200701  0.026511639 0.8769653
#> [24,] 0.56682698 0.8317866 0.70407139 0.71746907  0.016628721 0.8980712
#> [25,] 0.68921440 0.6727508 0.78713360 0.87749703  0.022613890 0.7934389
#>             V31        V32          V33         V34         V35       V36
#>  [1,] 0.6032804 0.59486604  0.485539463  0.34631966  0.17841069 0.1587811
#>  [2,] 0.4277574 0.34998356  0.241741067  0.12694833  0.12240793 0.1138537
#>  [3,] 0.9298783 0.77162903  0.404351525  0.38800923  0.48214525 0.1866981
#>  [4,] 0.2416983 0.31040420  0.354462262  0.49165614  0.56507827 0.2862287
#>  [5,] 0.2951233 0.09949002 -0.006007865  0.12928551  0.22367980 0.2371602
#>  [6,] 0.9221273 0.86602092  0.667961559  0.67080351  0.71561609 0.4109646
#>  [7,] 0.1648941 0.17825966  0.261660608  0.21966336  0.33132400 0.0566090
#>  [8,] 0.3731656 0.25531575  0.129563565  0.17259933  0.28160342 0.4262491
#>  [9,] 0.2865648 0.41043461  0.552996415  0.62021354  0.60160532 0.6323069
#> [10,] 0.3033974 0.52132640  0.470582900  0.17563964 -0.04194835 0.1727214
#> [11,] 0.9272867 0.88916426  0.723023557  0.65256213  0.66651385 0.4586679
#> [12,] 0.2866312 0.33845413  0.594140250  0.73489704  0.88661892 0.8808755
#> [13,] 0.1956420 0.11502536  0.278338919  0.33431971  0.38633460 0.4196911
#> [14,] 0.8512000 0.50450000  0.186200000  0.27090000  0.42320000 0.3043000
#> [15,] 0.4766000 0.24830000  0.537200000  0.62790000  0.36470000 0.4572000
#> [16,] 0.4645930 0.41946184  0.296124998 -0.01688673 -0.01853803 0.1309403
#> [17,] 0.5566958 0.32777212  0.271996558  0.17240529  0.38550254 0.5851806
#> [18,] 0.5126325 0.16842339  0.459700495  0.45395929  0.22186315 0.3394332
#> [19,] 0.5596804 0.59885787  0.661051320  0.61557267  0.61955745 0.7245126
#> [20,] 0.8028459 0.72313563  0.590051597  0.57374273  0.67072613 0.6219013
#> [21,] 0.2413144 0.21892642  0.240852106  0.25374214  0.24017798 0.2592204
#> [22,] 0.5672301 0.49540466  0.508454062  0.52933079  0.53040817 0.4782428
#> [23,] 0.7681568 0.68963538  0.737324636  0.84903769  0.96998031 0.9940955
#> [24,] 0.6043334 0.36039128  0.682210465  0.88800349  0.79283326 0.7551322
#> [25,] 0.6946578 0.64072406  0.662867915  0.71356805  0.67110321 0.5810005
#>             V37       V38        V39           V4          V40        V41
#>  [1,] 0.2384986 0.2766212 0.33533890  0.050610711  0.335825052 0.34972526
#>  [2,] 0.1435437 0.1591637 0.17279586  0.063954023  0.185798152 0.18127877
#>  [3,] 0.3571740 0.3732743 0.27989521  0.144341995  0.056976413 0.54074250
#>  [4,] 0.1405280 0.2688449 0.32008244  0.154391519  0.168118569 0.06350815
#>  [5,] 0.1407387 0.1627327 0.35086976  0.056315333  0.359013073 0.21736582
#>  [6,] 0.1500442 0.2108412 0.17312986  0.045802624  0.263405592 0.65644060
#>  [7,] 0.0511107 0.1777176 0.14572700  0.144470017 -0.007605258 0.08475899
#>  [8,] 0.3417211 0.3228016 0.37403487  0.386747428  0.412586686 0.35242872
#>  [9,] 0.5393980 0.3773135 0.33054509  0.027254987  0.295052546 0.25816574
#> [10,] 0.3621684 0.3581905 0.25544229  0.043142040  0.141737211 0.14226891
#> [11,] 0.2565332 0.4313365 0.49051453  0.088810293  0.419807290 0.57578802
#> [12,] 0.8201261 0.9147168 0.81947477  0.058921274  0.557141676 0.30929799
#> [13,] 0.4222164 0.5217492 0.43333583  0.094774409  0.314829453 0.35538809
#> [14,] 0.6116000 0.6756000 0.53750000  0.108300000  0.471900000 0.46470000
#> [15,] 0.6359000 0.6474000 0.55200000  0.080900000  0.325300000 0.22920000
#> [16,] 0.1594549 0.1742098 0.09472925 -0.002761169  0.063187610 0.04789943
#> [17,] 0.5002189 0.2266650 0.18012769  0.055623172  0.271443637 0.34438112
#> [18,] 0.4625838 0.5682464 0.65683965  0.009117630  0.523826369 0.21542757
#> [19,] 0.8059910 0.8810140 0.94872420  0.024207258  0.908038136 0.81411596
#> [20,] 0.5528104 0.5367551 0.33470710  0.032707542  0.180795960 0.17163178
#> [21,] 0.2766496 0.2193887 0.27673388 -0.003387410  0.278951755 0.25374870
#> [22,] 0.4451934 0.3427572 0.16784131  0.024790524  0.151827730 0.16399527
#> [23,] 0.9347068 0.8144988 0.64373217  0.032612036  0.463155383 0.29417872
#> [24,] 0.5720347 0.4622838 0.09976616  0.041963605  0.611565697 0.73721651
#> [25,] 0.4260636 0.2980357 0.25018094  0.024289160  0.237641680 0.21115286
#>              V42        V43          V44         V45          V46        V47
#>  [1,] 0.36625163 0.33839043  0.317043725 0.335242123  0.261655774 0.19486456
#>  [2,] 0.22115365 0.26243695  0.217848969 0.154639564  0.182419827 0.14915866
#>  [3,] 0.68779899 0.45064788  0.323673187 0.570132732  0.570407458 0.28693608
#>  [4,] 0.02761288 0.08802334  0.056670637 0.032330154  0.047561979 0.10423550
#>  [5,] 0.14124972 0.31904143  0.282421490 0.142317765  0.131566402 0.16272258
#>  [6,] 0.81936379 0.57905983  0.462726414 0.632825716  0.643278438 0.35246027
#>  [7,] 0.15715946 0.15603491  0.145227718 0.066920572  0.007781047 0.07550770
#>  [8,] 0.27450595 0.17566594  0.194826613 0.272568033  0.218961384 0.17979756
#>  [9,] 0.22590042 0.28892578  0.311541586 0.243411958  0.159365897 0.09769981
#> [10,] 0.12045149 0.04981566  0.082638988 0.060016319  0.006529779 0.02120127
#> [11,] 0.48730287 0.25073764  0.440803888 0.544564397  0.470914474 0.34075995
#> [12,] 0.16063632 0.13295037  0.250838532 0.462214095  0.361033059 0.20946676
#> [13,] 0.43024815 0.36939541  0.289458557 0.194722612  0.029048625 0.02957582
#> [14,] 0.25870000 0.21290000  0.222200000 0.211100000  0.017600000 0.13480000
#> [15,] 0.06530000 0.00000000  0.000000000 0.000000000  0.000000000 0.00000000
#> [16,] 0.14094006 0.15400261  0.137349643 0.011738009  0.020619570 0.05758575
#> [17,] 0.31383810 0.21327538  0.144582097 0.158877258  0.159767423 0.10783888
#> [18,] 0.21091869 0.16745992  0.060749147 0.004995938 -0.013046501 0.06629022
#> [19,] 0.71919530 0.53672350  0.339899644 0.437033016  0.475513616 0.36768713
#> [20,] 0.18543285 0.23153749  0.198943325 0.150124865  0.122720512 0.06962439
#> [21,] 0.22082199 0.18421378  0.179096509 0.133255700  0.098614168 0.02655217
#> [22,] 0.09284115 0.02818856 -0.001569024 0.034742471  0.020119934 0.02038464
#> [23,] 0.19055728 0.15722067  0.108909029 0.177556882  0.193119566 0.11035699
#> [24,] 0.60497209 0.34122035  0.174131047 0.240240465  0.348476511 0.23346767
#> [25,] 0.22099244 0.22788307  0.217857337 0.143370990  0.095871328 0.06301551
#>              V48          V49         V5         V50          V51         V52
#>  [1,] 0.16279535  0.086828453 0.09645521 0.027101694 0.0220330228 0.016998441
#>  [2,] 0.08740218  0.050835903 0.08575040 0.017011709 0.0161768818 0.015500557
#>  [3,] 0.14011258  0.109820568 0.15442696 0.043053744 0.0363021307 0.018725185
#>  [4,] 0.11057328  0.076282314 0.14720935 0.026972691 0.0221923329 0.034059705
#>  [5,] 0.14162644  0.110718016 0.06392080 0.033373443 0.0374202145 0.005952286
#>  [6,] 0.17767898  0.061647474 0.09219366 0.037935448 0.0567218778 0.048821825
#>  [7,] 0.07545966  0.012776059 0.13185629 0.009324588 0.0310293935 0.014025311
#>  [8,] 0.14470312  0.066226613 0.36238402 0.018844354 0.0147456642 0.034252114
#>  [9,] 0.07390871  0.037813502 0.03084500 0.013457851 0.0093300103 0.007612973
#> [10,] 0.03933516  0.011390358 0.05155710 0.003658401 0.0128212385 0.007618902
#> [11,] 0.19483923  0.101250762 0.04805769 0.035379229 0.0289425732 0.027792290
#> [12,] 0.17966668  0.096545779 0.03523125 0.040142116 0.0262066394 0.013672414
#> [13,] 0.02299812  0.018832869 0.08744140 0.008710651 0.0172396150 0.011335848
#> [14,] 0.07440000  0.013000000 0.09740000 0.010600000 0.0033000000 0.023200000
#> [15,] 0.00000000  0.000000000 0.05210000 0.000000000 0.0000000000 0.005600000
#> [16,] 0.04990013 -0.001014932 0.01511808 0.017912380 0.0058891785 0.004196904
#> [17,] 0.09463004  0.053270600 0.05891102 0.020176486 0.0116483605 0.011750500
#> [18,] 0.07611167  0.019477623 0.04121771 0.011799639 0.0062376331 0.022377850
#> [19,] 0.24171824  0.155608397 0.05656562 0.060825090 0.0341662181 0.014524882
#> [20,] 0.04460309  0.029371442 0.05469574 0.020108394 0.0041319344 0.010935563
#> [21,] 0.02780661  0.029376739 0.01246885 0.006871724 0.0005401322 0.005711345
#> [22,] 0.02818923  0.005617010 0.04237521 0.007400901 0.0041183509 0.005643176
#> [23,] 0.05974963  0.017478494 0.02602858 0.024825615 0.0246788907 0.018235253
#> [24,] 0.06263244  0.018235117 0.06732151 0.048323605 0.0435737209 0.036233139
#> [25,] 0.04908888  0.024367101 0.04655151 0.014614384 0.0116734532 0.007996809
#>                V53         V54         V55          V56          V57
#>  [1,]  0.011616406 0.010496240 0.011578047 0.0097964249 0.0075107336
#>  [2,]  0.008794303 0.010456250 0.002722564 0.0058058608 0.0055845594
#>  [3,]  0.018932909 0.015446138 0.012700000 0.0141977508 0.0127601953
#>  [4,]  0.012082278 0.011777711 0.018622795 0.0116881698 0.0031231306
#>  [5,]  0.016749871 0.012496147 0.037400257 0.0119858595 0.0122462338
#>  [6,]  0.028272702 0.018386377 0.014678992 0.0120475933 0.0051863774
#>  [7,]  0.031364156 0.012222085 0.016848773 0.0202759467 0.0211603366
#>  [8,]  0.020681048 0.015468347 0.019253026 0.0021771181 0.0121491956
#>  [9,]  0.002947562 0.009132133 0.005516507 0.0058305459 0.0038660842
#> [10,] -0.004280732 0.004917195 0.001907822 0.0045051608 0.0060154830
#> [11,]  0.010379446 0.020696663 0.025576593 0.0129480034 0.0106513208
#> [12,]  0.009205386 0.015764499 0.018390547 0.0133936446 0.0173318823
#> [13,]  0.015780796 0.021594118 0.008861782 0.0189885213 0.0149678554
#> [14,]  0.016600000 0.009500000 0.018000000 0.0244000000 0.0316000000
#> [15,]  0.023700000 0.020400000 0.005000000 0.0137000000 0.0164000000
#> [16,]  0.023223560 0.020496309 0.010151870 0.0063860002 0.0018138366
#> [17,]  0.008341654 0.010872705 0.009693634 0.0085710963 0.0110735543
#> [18,]  0.028315558 0.016134840 0.025479787 0.0029192820 0.0006322569
#> [19,]  0.012452816 0.010822632 0.017764624 0.0136268856 0.0023516228
#> [20,]  0.001777027 0.005501104 0.007985989 0.0035523825 0.0103028085
#> [21,]  0.004016197 0.008548823 0.005385483 0.0009228006 0.0036749772
#> [22,]  0.012352018 0.006199879 0.004295941 0.0064139730 0.0019042789
#> [23,]  0.007909241 0.007957217 0.009086024 0.0074503130 0.0078505423
#> [24,]  0.015978721 0.015356860 0.011183488 0.0043624419 0.0088898838
#> [25,]  0.008870051 0.008679019 0.007207042 0.0067906561 0.0086606056
#>                V58           V59          V6          V60         V7         V8
#>  [1,] 0.0085779740  7.486433e-03 0.116035004 0.0051354012 0.12503194 0.12464721
#>  [2,] 0.0091566258  5.247862e-03 0.118485978 0.0060031017 0.14364099 0.18454425
#>  [3,] 0.0144353063  1.460450e-02 0.088377508 0.0199106837 0.08672817 0.11543986
#>  [4,] 0.0112919986  2.319375e-02 0.041522200 0.0119799037 0.07664672 0.05168121
#>  [5,] 0.0310761715  1.328896e-02 0.152247558 0.0038751768 0.29332777 0.33669046
#>  [6,] 0.0130310167  1.050520e-02 0.220714177 0.0083682711 0.16168693 0.19711176
#>  [7,] 0.0153441947  1.341177e-02 0.102984328 0.0160060915 0.09681975 0.06222434
#>  [8,] 0.0138049410  1.793669e-02 0.168620148 0.0215621956 0.17787297 0.03570931
#>  [9,] 0.0078193220  7.621640e-03 0.058536934 0.0038897069 0.10550813 0.16769191
#> [10,] 0.0050574678 -3.391630e-06 0.089606911 0.0008933656 0.15172564 0.14458450
#> [11,] 0.0091477381  2.331371e-02 0.103076608 0.0127325402 0.08333884 0.08535518
#> [12,] 0.0142369199  1.012846e-02 0.123512008 0.0107843982 0.25893622 0.33312285
#> [13,] 0.0134856544  2.654285e-02 0.103816139 0.0172835228 0.16791943 0.10168530
#> [14,] 0.0164000000  9.500000e-03 0.228000000 0.0078000000 0.24310000 0.37710000
#> [15,] 0.0081000000  1.390000e-02 0.105100000 0.0111000000 0.01450000 0.06740000
#> [16,] 0.0028190071  4.422551e-03 0.057106997 0.0005839498 0.10151944 0.02193475
#> [17,] 0.0063280857  7.643174e-03 0.110923433 0.0048898753 0.15736543 0.18069666
#> [18,] 0.0047740502  8.951900e-03 0.048441793 0.0090319718 0.03285627 0.06922143
#> [19,] 0.0136339152  9.972726e-03 0.094000700 0.0109141807 0.19045934 0.21112762
#> [20,] 0.0111960872  7.899716e-03 0.113120005 0.0085750710 0.12873266 0.13279628
#> [21,] 0.0062017479  3.689971e-03 0.004071616 0.0031794009 0.03470975 0.05849867
#> [22,] 0.0027982264  4.272137e-03 0.068065016 0.0048136971 0.07794900 0.07247934
#> [23,] 0.0099718191  6.243012e-03 0.099362913 0.0102709638 0.14133052 0.08207451
#> [24,] 0.0007296512  5.703837e-03 0.114044303 0.0044074420 0.16500907 0.15886384
#> [25,] 0.0046731000  8.090937e-03 0.073160763 0.0055386002 0.08294972 0.08598500
#>               V9
#>  [1,] 0.22020256
#>  [2,] 0.25831733
#>  [3,] 0.17391172
#>  [4,] 0.10844657
#>  [5,] 0.31433478
#>  [6,] 0.23244290
#>  [7,] 0.22059895
#>  [8,] 0.18466420
#>  [9,] 0.30574872
#> [10,] 0.22899605
#> [11,] 0.24793442
#> [12,] 0.49433426
#> [13,] 0.02915636
#> [14,] 0.55980000
#> [15,] 0.12940000
#> [16,] 0.03704273
#> [17,] 0.20249243
#> [18,] 0.14188896
#> [19,] 0.01691858
#> [20,] 0.13301489
#> [21,] 0.07022584
#> [22,] 0.12589613
#> [23,] 0.07290645
#> [24,] 0.14795581
#> [25,] 0.10326753
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986