Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12       V13       V14
#>  [1,] 0.101843266 0.59589634 0.47957974 0.24135919 0.1169372 0.3240865
#>  [2,] 0.016438007 0.19619388 0.26459741 0.32394777 0.3429315 0.3717936
#>  [3,] 0.028161043 0.29566208 0.37153045 0.45387993 0.4303565 0.4054056
#>  [4,] 0.026603233 0.20128000 0.25572301 0.28398005 0.2619085 0.2118727
#>  [5,] 0.014342683 0.28933912 0.31574797 0.30745331 0.2158103 0.1188685
#>  [6,] 0.017812022 0.25286774 0.29635246 0.34133139 0.3211361 0.2555964
#>  [7,] 0.024114804 0.21456505 0.22134233 0.23599063 0.3089072 0.3857165
#>  [8,] 0.077621231 0.50368496 0.34577177 0.15099678 0.2279637 0.3277910
#>  [9,] 0.044673056 0.01793408 0.15361679 0.29622269 0.3345605 0.4192030
#> [10,] 0.032784621 0.25835422 0.29058769 0.27208323 0.3186169 0.4071190
#> [11,] 0.051156031 0.18723378 0.21477158 0.25380549 0.2896947 0.2526089
#> [12,] 0.114626871 0.35024702 0.44835312 0.53076936 0.6634072 0.7667300
#> [13,] 0.033268600 0.26051777 0.31694612 0.31553150 0.2689211 0.1336260
#> [14,] 0.009583776 0.09913177 0.12730426 0.12420352 0.1681792 0.1820702
#> [15,] 0.021702564 0.12832994 0.17386788 0.28263277 0.2203446 0.1879341
#> [16,] 0.032195848 0.31096462 0.23125602 0.20644003 0.2375782 0.1543514
#> [17,] 0.027638856 0.07332460 0.09588428 0.16981992 0.1248239 0.1150824
#> [18,] 0.008000242 0.09404447 0.09692343 0.12281154 0.1520205 0.1554536
#> [19,] 0.018189313 0.10526591 0.08181297 0.10885233 0.2448012 0.2886865
#> [20,] 0.012687632 0.09894315 0.11551403 0.09794144 0.1424599 0.2113246
#> [21,] 0.019667412 0.09480486 0.09118501 0.07301178 0.1792377 0.3146484
#> [22,] 0.020089970 0.09949572 0.13694834 0.17885107 0.1862567 0.1892226
#> [23,] 0.018525903 0.16843289 0.20894348 0.09126959 0.1744563 0.3866333
#> [24,] 0.030165965 0.35944052 0.26642221 0.06522157 0.1481080 0.4817297
#> [25,] 0.023283747 0.29088781 0.28845352 0.38408790 0.4017767 0.2914621
#>             V15       V16       V17       V18       V19           V2       V20
#>  [1,] 0.4666777 0.4640324 0.1870456 0.1584593 0.3739769  0.114470150 0.9007700
#>  [2,] 0.3891851 0.4815611 0.5623282 0.6637419 0.7271267  0.014352253 0.8214946
#>  [3,] 0.3182602 0.3300020 0.2997682 0.2986871 0.3192303  0.033867142 0.6134680
#>  [4,] 0.1734054 0.1815885 0.2186872 0.2730373 0.3446013  0.034796497 0.3977451
#>  [5,] 0.1726035 0.3131803 0.3607246 0.4427077 0.5110504  0.022633645 0.5847568
#>  [6,] 0.1946513 0.2149755 0.2400390 0.3474012 0.5492116  0.027936541 0.6705395
#>  [7,] 0.4874827 0.5865062 0.6669031 0.7037767 0.7275244  0.018379129 0.7705367
#>  [8,] 0.4124156 0.4986356 0.2175090 0.1769673 0.2076966  0.077491829 0.8511252
#>  [9,] 0.5267898 0.6105108 0.7589276 0.8513626 0.9058917  0.059501509 0.9132790
#> [10,] 0.5390311 0.6451807 0.7742545 0.7528216 0.7583424  0.054672864 0.8022385
#> [11,] 0.2418190 0.2663991 0.2807290 0.1750889 0.2296336  0.066503237 0.3020448
#> [12,] 0.9702052 1.0169432 0.8476307 0.8759898 0.9304388  0.128107964 0.9811947
#> [13,] 0.1293091 0.1699096 0.2005361 0.2510246 0.2136570  0.023963572 0.2042667
#> [14,] 0.1698443 0.2120749 0.2890591 0.3388011 0.3930134  0.005973975 0.4414349
#> [15,] 0.1310341 0.1261877 0.2103342 0.4366983 0.3902294  0.044431361 0.5940130
#> [16,] 0.3698876 0.3348583 0.2312477 0.2205921 0.2562526 -0.001844243 0.4095917
#> [17,] 0.1253499 0.1916347 0.1844558 0.2397836 0.3337511  0.041115296 0.2799117
#> [18,] 0.1803533 0.1779584 0.1512304 0.1722161 0.1440162  0.010562942 0.1539858
#> [19,] 0.2634652 0.2776785 0.2180937 0.2963202 0.6221779  0.031851066 0.8349603
#> [20,] 0.2319652 0.2117361 0.1293733 0.1505908 0.1705469  0.013503247 0.2035814
#> [21,] 0.5209569 0.7100572 0.7672544 0.7640346 0.8293273  0.018659739 0.8638939
#> [22,] 0.2345714 0.4279412 0.5440895 0.4684414 0.3590544  0.020682476 0.3824317
#> [23,] 0.5006503 0.7094793 0.9039347 0.9494278 0.7851589  0.011454929 0.4999022
#> [24,] 0.6336325 0.6712029 0.7202426 0.6057138 0.3748721  0.055022856 0.3326985
#> [25,] 0.3006945 0.4414518 0.7200553 0.5882870 0.2012211  0.025785681 0.2949466
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.9458372 0.4583567 0.1268029 0.1287445 0.2639186 0.5228948 0.3328341
#>  [2,] 0.9194608 0.9922858 0.9478545 0.9130553 0.9394963 0.8836818 0.8199701
#>  [3,] 0.7288719 0.6818707 0.6005086 0.6688814 0.6506487 0.8517173 0.9915176
#>  [4,] 0.4615862 0.5109543 0.5986133 0.6963420 0.7808577 0.8425008 0.8901581
#>  [5,] 0.6933787 0.8277629 0.9069844 0.9471983 0.9056275 0.7708004 0.5729768
#>  [6,] 0.7792162 0.8031045 0.8255058 0.9245245 0.9399101 0.9663445 0.9826765
#>  [7,] 0.8536030 0.8979875 0.9228907 0.9273408 0.8993801 0.9047225 0.9140973
#>  [8,] 0.9951504 0.3891673 0.1175926 0.1916150 0.2082384 0.3713855 0.2906166
#>  [9,] 0.8602746 0.8670936 0.8941078 0.9260850 0.8127042 0.6209593 0.4888176
#> [10,] 0.8586899 0.7617359 0.5741541 0.3791336 0.2268688 0.2887490 0.3976521
#> [11,] 0.3396881 0.4226296 0.5164148 0.5190788 0.5359622 0.5561333 0.6524740
#> [12,] 0.8850483 0.8089782 0.9214454 0.8696648 0.6013067 0.5157839 0.5365456
#> [13,] 0.3282852 0.6094311 0.8108354 0.8740492 0.9433630 0.9604460 0.9512510
#> [14,] 0.5224171 0.6004765 0.6761680 0.7736815 0.8917404 0.8991544 0.9307096
#> [15,] 0.7156595 0.7956898 0.8391012 0.8367393 0.8270327 0.7134708 0.5467065
#> [16,] 0.3021859 0.6277992 0.5654799 0.3027599 0.2159017 0.5502321 0.8069103
#> [17,] 0.2731903 0.1857982 0.2741616 0.2957608 0.1847959 0.2901649 0.3856502
#> [18,] 0.2301002 0.2190151 0.3603873 0.5985334 0.8647240 1.0000000 0.8362344
#> [19,] 0.7714084 0.6986229 0.4887081 0.4645026 0.6623702 0.8146606 0.8970277
#> [20,] 0.2383901 0.3597579 0.5488864 0.6713826 0.7058803 0.7651741 0.7709159
#> [21,] 0.9721221 0.9638063 0.9396852 0.9424072 0.8941770 0.9276886 0.8884366
#> [22,] 0.4116141 0.4581624 0.5316042 0.6394073 0.7244929 0.7477575 0.7491343
#> [23,] 0.5552215 0.7314649 0.8074697 0.8381163 0.7384619 0.6661999 0.5751563
#> [24,] 0.5453994 0.8054316 0.8863650 0.9053142 0.7985481 0.7182481 0.7100050
#> [25,] 0.7143213 0.7004315 0.3784202 0.3099372 0.5558089 0.5388179 0.5401189
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.5930765 0.5803363 0.079200116 0.8437115 0.5331717 0.5045672 0.52424098
#>  [2,] 0.7355300 0.6399576 0.016772629 0.5097943 0.3212326 0.2397414 0.21653665
#>  [3,] 0.9406409 0.8277509 0.067027790 0.6536037 0.5647843 0.7024233 0.57678962
#>  [4,] 0.8971232 0.8449930 0.041774104 0.7125650 0.6103368 0.4555737 0.38492373
#>  [5,] 0.3510233 0.1397824 0.023506106 0.2762325 0.3906403 0.4024923 0.49330951
#>  [6,] 0.9410121 0.8208597 0.033142301 0.6148994 0.4190333 0.3251574 0.25888310
#>  [7,] 0.8980796 0.7590974 0.024817958 0.5831369 0.4289613 0.3671659 0.24908566
#>  [8,] 0.3990156 0.3949697 0.074025441 0.5678425 0.2239855 0.3215407 0.50950139
#>  [9,] 0.4137504 0.3432335 0.078618527 0.2376859 0.2580458 0.3561975 0.43267528
#> [10,] 0.3035645 0.1532003 0.051974860 0.3357624 0.2307384 0.3968161 0.43082392
#> [11,] 0.7999421 0.8039590 0.067772369 0.8086816 0.8207498 0.7395516 0.57245639
#> [12,] 0.6108616 0.5646540 0.085243449 0.4422140 0.3313488 0.2510816 0.08392838
#> [13,] 0.9996960 0.8676574 0.029390191 0.6831119 0.7278356 0.6845492 0.50780477
#> [14,] 0.8759297 0.7161492 0.008722252 0.5901580 0.4792832 0.3945143 0.31007098
#> [15,] 0.4930809 0.5205304 0.056953392 0.5325505 0.6277810 0.6346883 0.72458749
#> [16,] 0.6759583 0.7232281 0.010502674 0.8978913 0.5884239 0.3476534 0.72011927
#> [17,] 0.5320487 0.6786317 0.041813865 0.7460078 0.7511871 0.5661179 0.60495755
#> [18,] 0.6494807 0.5486228 0.010197729 0.3861555 0.3108412 0.1789474 0.17713639
#> [19,] 0.7351298 0.5142229 0.027720862 0.4109639 0.3333099 0.2172856 0.28184059
#> [20,] 0.8592060 0.9156026 0.023752273 0.7941395 0.6382416 0.5289560 0.56486988
#> [21,] 0.6912864 0.4469479 0.026629404 0.4149771 0.4273988 0.3419676 0.30934978
#> [22,] 0.8155474 0.8141538 0.022795803 0.7675373 0.8025758 0.7268611 0.64530285
#> [23,] 0.4149228 0.3615548 0.045498323 0.4410568 0.2431357 0.1744110 0.18073169
#> [24,] 0.5904829 0.4123481 0.081855956 0.4743521 0.5046368 0.2787993 0.22356718
#> [25,] 0.9035855 0.8314041 0.008423674 0.4200732 0.2725011 0.2417482 0.09395772
#>             V34        V35        V36       V37        V38        V39
#>  [1,] 0.6670871 0.83723391 0.81183927 0.8101578 0.89053539 0.76868539
#>  [2,] 0.1338239 0.16005310 0.22120432 0.2653208 0.29875283 0.39878907
#>  [3,] 0.2745270 0.03687205 0.24336591 0.3174326 0.44415660 0.58757691
#>  [4,] 0.2974138 0.19021823 0.14137719 0.2156214 0.19717336 0.19865073
#>  [5,] 0.5882576 0.51617710 0.35653430 0.2439421 0.19869702 0.22906716
#>  [6,] 0.2151808 0.17941008 0.13465203 0.1369922 0.14757624 0.15008205
#>  [7,] 0.1131566 0.08412157 0.14236356 0.2325173 0.22383271 0.18731141
#>  [8,] 0.6678855 0.71254407 0.78149706 0.7569073 0.81624092 0.67732977
#>  [9,] 0.4312548 0.41227325 0.33762842 0.3843667 0.43066092 0.43606479
#> [10,] 0.3120912 0.31678220 0.28847117 0.2934529 0.39873556 0.29785277
#> [11,] 0.5546526 0.57973358 0.44888635 0.3229553 0.38922049 0.37725221
#> [12,] 0.1435998 0.19788938 0.12524254 0.1708436 0.15969415 0.19653666
#> [13,] 0.1992947 0.09588789 0.10871650 0.1214216 0.14181615 0.18038056
#> [14,] 0.2026144 0.19117710 0.25673206 0.2529441 0.26540686 0.23593752
#> [15,] 0.6601158 0.61091478 0.54566192 0.4284376 0.27468147 0.16178784
#> [16,] 0.9087795 0.79443490 0.76083645 0.5774488 0.46695477 0.07551621
#> [17,] 0.7208722 0.76053030 0.74991004 0.6661062 0.67266543 0.68729102
#> [18,] 0.2061991 0.28412855 0.21449467 0.0701367 0.08370281 0.13184735
#> [19,] 0.3100161 0.28986807 0.30370853 0.4140633 0.47255183 0.48675953
#> [20,] 0.5795399 0.49211958 0.37895119 0.1978410 0.16654915 0.20533409
#> [21,] 0.2805005 0.17952938 0.08063424 0.1855062 0.14634512 0.09976259
#> [22,] 0.6775019 0.71689754 0.69469672 0.6213445 0.46511458 0.35580750
#> [23,] 0.1068044 0.37619008 0.52977343 0.4211587 0.21440887 0.10346739
#> [24,] 0.1297602 0.42752217 0.65732529 0.6837005 0.31605974 0.23883127
#> [25,] 0.1480895 0.28404420 0.17188060 0.2054001 0.22052584 0.09548041
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.13334486 0.51509315 0.39177614 0.45296963 0.45607877 0.43114789
#>  [2,] 0.06943282 0.39097894 0.32613591 0.28764201 0.27745552 0.21701907
#>  [3,] 0.03060236 0.44407196 0.40783112 0.60497261 0.54816561 0.53255676
#>  [4,] 0.05495242 0.21274834 0.22920034 0.28872821 0.23041416 0.14807699
#>  [5,] 0.02816645 0.33054432 0.28773029 0.27360118 0.26455879 0.19783360
#>  [6,] 0.02987893 0.14487920 0.07266704 0.11504214 0.20680558 0.23338830
#>  [7,] 0.05193072 0.23241197 0.23345071 0.23080956 0.20558458 0.17314251
#>  [8,] 0.07385648 0.50784036 0.32888676 0.20418047 0.19176202 0.20594005
#>  [9,] 0.05873144 0.44766397 0.38235329 0.38142604 0.27620251 0.23963265
#> [10,] 0.06340259 0.14596250 0.05008494 0.17884487 0.19794211 0.12034686
#> [11,] 0.06757756 0.32875753 0.54139416 0.58652928 0.47548570 0.44196100
#> [12,] 0.01595190 0.25156718 0.15427700 0.15435383 0.17449195 0.24624540
#> [13,] 0.05163985 0.05804825 0.15264446 0.07217551 0.10556396 0.15286333
#> [14,] 0.01521601 0.17064490 0.14212469 0.16566383 0.15283404 0.13538087
#> [15,] 0.05929530 0.39763697 0.40958231 0.43664184 0.40446978 0.26627686
#> [16,] 0.04184891 0.62345841 0.75463504 0.60027189 0.32296543 0.17488675
#> [17,] 0.01519360 0.64882832 0.52759333 0.47921031 0.36076688 0.29645981
#> [18,] 0.03419743 0.20999839 0.27548447 0.26126570 0.19622779 0.08835313
#> [19,] 0.02531506 0.34618825 0.19833145 0.16298622 0.10561332 0.07372710
#> [20,] 0.02900512 0.18574713 0.17394726 0.20032384 0.24553359 0.16383463
#> [21,] 0.04135726 0.12726986 0.09578068 0.03970379 0.06837913 0.11970442
#> [22,] 0.03208485 0.29496822 0.22415150 0.18161856 0.14687422 0.14884602
#> [23,] 0.03906465 0.29814081 0.38867637 0.28982372 0.16911331 0.17163513
#> [24,] 0.05834961 0.47635160 0.48158163 0.37274292 0.24400787 0.16733740
#> [25,] 0.03659501 0.03952950 0.15884615 0.16816018 0.18435743 0.16826493
#>              V45         V46        V47        V48         V49           V5
#>  [1,] 0.54005280 0.419621349 0.09036798 0.09511101 0.039142407  0.128860616
#>  [2,] 0.19839355 0.149596857 0.12093612 0.08647683 0.062775794  0.114799530
#>  [3,] 0.55127852 0.294844106 0.20663362 0.21497269 0.115134531  0.093912397
#>  [4,] 0.12728179 0.136342461 0.15168326 0.11779728 0.057236364  0.067876352
#>  [5,] 0.17357765 0.149695841 0.13751383 0.11090830 0.058240878  0.041331081
#>  [6,] 0.21267040 0.208052737 0.15786906 0.09246456 0.046075113  0.062328482
#>  [7,] 0.12009268 0.109845150 0.12780782 0.08914636 0.046078578  0.090296770
#>  [8,] 0.38788982 0.279029030 0.14101916 0.16145184 0.107279944  0.057004054
#>  [9,] 0.17831769 0.131532489 0.06735948 0.03164123 0.033054935  0.141575382
#> [10,] 0.05095142 0.006228314 0.04744895 0.06984572 0.044553885  0.046226066
#> [11,] 0.52393825 0.516449165 0.35681593 0.20815901 0.109456984  0.064094354
#> [12,] 0.13593845 0.067441880 0.08428703 0.08532466 0.071497235  0.030788729
#> [13,] 0.08858135 0.041388685 0.09041240 0.11869618 0.066770843  0.070365925
#> [14,] 0.10011442 0.062268140 0.05635377 0.05130398 0.019733277  0.039755140
#> [15,] 0.12935669 0.091979212 0.06694232 0.06150533 0.024827270  0.112393787
#> [16,] 0.21426655 0.331639410 0.22934328 0.06328794 0.010549563  0.074378547
#> [17,] 0.22569139 0.151319228 0.14114390 0.10114371 0.054649717  0.044876089
#> [18,] 0.08867418 0.095092946 0.03963497 0.03861369 0.031783621  0.055046624
#> [19,] 0.10644585 0.099600692 0.06921179 0.03783326 0.017658414  0.040221365
#> [20,] 0.06021175 0.050419998 0.04868949 0.03952885 0.021293335  0.033968081
#> [21,] 0.10090765 0.036306078 0.05697096 0.04529342 0.026525951  0.062222643
#> [22,] 0.13134173 0.126994356 0.07654415 0.05430720 0.027731804  0.053765464
#> [23,] 0.15876924 0.057926017 0.10309128 0.11534803 0.066977844  0.094231616
#> [24,] 0.18061094 0.169912698 0.16495840 0.15280742 0.092027846  0.049519350
#> [25,] 0.08424505 0.054485203 0.01148158 0.02791309 0.009879682 -0.005192288
#>               V50          V51         V52         V53         V54         V55
#>  [1,] 0.022753595  0.015237272 0.041834875 0.025432020 0.013149086 0.013590582
#>  [2,] 0.027596708  0.017917477 0.012014605 0.014825836 0.013391726 0.009879908
#>  [3,] 0.028228199  0.035128205 0.019174188 0.004240162 0.010707339 0.013515605
#>  [4,] 0.014157560  0.017106818 0.013982782 0.009390206 0.009056545 0.006719860
#>  [5,] 0.018486325  0.012274764 0.008432271 0.008454076 0.011368073 0.006192398
#>  [6,] 0.014771326  0.017504724 0.012348878 0.007091776 0.005954138 0.004140661
#>  [7,] 0.017480407  0.013309637 0.007414474 0.010065375 0.010628619 0.004039333
#>  [8,] 0.043360786  0.026612566 0.010832931 0.010162705 0.031375240 0.018297803
#>  [9,] 0.006379924 -0.001218478 0.011488700 0.014909173 0.012261540 0.010853411
#> [10,] 0.021743661  0.020069321 0.018396051 0.009583350 0.014707225 0.009835632
#> [11,] 0.041203484  0.045062214 0.036402483 0.016957313 0.017376240 0.016783157
#> [12,] 0.012236230  0.029560391 0.019441911 0.016037980 0.020396000 0.018780385
#> [13,] 0.021947998  0.010484167 0.009071374 0.007385259 0.011211458 0.005558594
#> [14,] 0.010860971  0.002971027 0.005523299 0.006312597 0.006827349 0.006153690
#> [15,] 0.014979670  0.007876352 0.005884588 0.014425757 0.011186380 0.008266948
#> [16,] 0.045581472  0.044764402 0.035451281 0.016343218 0.016176099 0.009348197
#> [17,] 0.017261945  0.010592702 0.009249063 0.015397430 0.012700436 0.009431081
#> [18,] 0.006523312  0.009460551 0.010748774 0.005704662 0.004676809 0.004904723
#> [19,] 0.021949968  0.017123745 0.013095192 0.020275096 0.020976023 0.012158828
#> [20,] 0.012096686  0.014306811 0.011387691 0.011681647 0.004822708 0.005276676
#> [21,] 0.013676883  0.015298071 0.007911930 0.011292698 0.008051329 0.004143595
#> [22,] 0.016686246  0.010910159 0.007748656 0.005289868 0.008092462 0.011094589
#> [23,] 0.005733337  0.023484277 0.008816698 0.008647559 0.008024167 0.005095779
#> [24,] 0.019656727  0.021580528 0.011129934 0.009710712 0.011422355 0.009048095
#> [25,] 0.012155895  0.009961996 0.001775791 0.008132761 0.013423412 0.017530634
#>               V56         V57          V58         V59          V6         V60
#>  [1,] 0.018614269 0.011187100 0.0174184936 0.014228254 0.137966158 0.016692352
#>  [2,] 0.008948675 0.007082122 0.0079762591 0.009861818 0.134066758 0.006289799
#>  [3,] 0.005309539 0.006620564 0.0026341553 0.004072895 0.097554619 0.002027062
#>  [4,] 0.006124308 0.005732620 0.0061154725 0.005178757 0.105072851 0.003630524
#>  [5,] 0.006950430 0.003456757 0.0063375594 0.002929659 0.066851617 0.004255573
#>  [6,] 0.004395081 0.003585168 0.0034642295 0.003535781 0.081759837 0.004328252
#>  [7,] 0.010516150 0.006297743 0.0129272423 0.007161818 0.124201813 0.006177268
#>  [8,] 0.023701579 0.014245775 0.0193758588 0.021768028 0.121138873 0.006579635
#>  [9,] 0.015815785 0.008802989 0.0075855772 0.012073441 0.151536818 0.007798632
#> [10,] 0.013544485 0.004671609 0.0136214103 0.006548352 0.001450038 0.011226795
#> [11,] 0.010908029 0.011501394 0.0116689348 0.014089675 0.114854002 0.008245118
#> [12,] 0.004900378 0.003057292 0.0124339796 0.011454140 0.138285209 0.014909874
#> [13,] 0.007835542 0.006020157 0.0064590495 0.008146265 0.110539066 0.007797309
#> [14,] 0.004116937 0.004084796 0.0048555123 0.004211003 0.053261624 0.003870649
#> [15,] 0.011198624 0.013179235 0.0080749947 0.004627410 0.170600609 0.012465940
#> [16,] 0.004065600 0.009556039 0.0002842912 0.006593723 0.113732882 0.006208690
#> [17,] 0.008311078 0.006711863 0.0112846025 0.010416348 0.072577725 0.004569392
#> [18,] 0.001423312 0.002055828 0.0041325160 0.003039570 0.104759645 0.001779080
#> [19,] 0.010488316 0.005877451 0.0104076768 0.002911077 0.028202932 0.004660434
#> [20,] 0.005178432 0.008935046 0.0047604019 0.006761328 0.065430169 0.006699043
#> [21,] 0.007257233 0.004116938 0.0030533525 0.003704268 0.080616975 0.002722271
#> [22,] 0.006666558 0.010929101 0.0078683495 0.010051489 0.110328972 0.007602436
#> [23,] 0.007490014 0.008914227 0.0140784258 0.019160962 0.103895331 0.015155905
#> [24,] 0.006511431 0.013520537 0.0091359568 0.007578997 0.100850804 0.007266740
#> [25,] 0.012353986 0.009343551 0.0081555318 0.005515097 0.070379845 0.006952575
#>               V7         V8         V9
#>  [1,] 0.16732323 0.38922192 0.65562672
#>  [2,] 0.14567683 0.11429532 0.15199399
#>  [3,] 0.09699537 0.15346012 0.25069773
#>  [4,] 0.12040744 0.12165299 0.16874328
#>  [5,] 0.10028472 0.11849090 0.22162413
#>  [6,] 0.10917941 0.12713498 0.20046131
#>  [7,] 0.12462692 0.13741028 0.19265787
#>  [8,] 0.12843463 0.42982668 0.59289852
#>  [9,] 0.13022123 0.04646039 0.05894813
#> [10,] 0.04613610 0.05908591 0.18654447
#> [11,] 0.10139488 0.14976920 0.19357917
#> [12,] 0.17387834 0.15068228 0.23362669
#> [13,] 0.12296681 0.13584757 0.15933343
#> [14,] 0.07932035 0.06053001 0.08160259
#> [15,] 0.16048653 0.14315661 0.08478414
#> [16,] 0.17462156 0.14325420 0.14564948
#> [17,] 0.07161337 0.09654816 0.07833589
#> [18,] 0.10815725 0.08929694 0.11750430
#> [19,] 0.04261018 0.05028090 0.09393817
#> [20,] 0.06152186 0.07891589 0.09693744
#> [21,] 0.13290463 0.09720809 0.10880565
#> [22,] 0.12467069 0.11417531 0.07631321
#> [23,] 0.17982009 0.17585542 0.13201619
#> [24,] 0.10598208 0.21928173 0.27585189
#> [25,] 0.13717887 0.17264763 0.21197930
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841