Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.017121476 0.16965997 0.24672491 0.27389417 0.29697358 0.28118909
#>  [2,] 0.020446809 0.17556614 0.11518746 0.04914957 0.09649885 0.23230132
#>  [3,] 0.030675425 0.22664251 0.26736028 0.31185272 0.32247172 0.35181592
#>  [4,] 0.028129836 0.25956776 0.31078195 0.34698088 0.29186808 0.19394352
#>  [5,] 0.036837417 0.22103039 0.25282049 0.26706960 0.31844210 0.29503213
#>  [6,] 0.023956638 0.37342193 0.46114620 0.51120965 0.49916551 0.40114430
#>  [7,] 0.028896130 0.27271604 0.30890496 0.28611100 0.22276972 0.09998104
#>  [8,] 0.053906222 0.25482322 0.38575741 0.46827881 0.53362033 0.48946057
#>  [9,] 0.063914765 0.37163414 0.34217191 0.20807191 0.18305379 0.24316656
#> [10,] 0.026124061 0.07996179 0.10227184 0.08443838 0.05995201 0.17468553
#> [11,] 0.030202351 0.25255726 0.26657043 0.26826594 0.26834985 0.27896123
#> [12,] 0.044334299 0.02937280 0.14147551 0.26950586 0.30127438 0.31902736
#> [13,] 0.010268860 0.16188796 0.20765944 0.30581315 0.40469295 0.50729322
#> [14,] 0.012731651 0.10289113 0.06129395 0.08226943 0.25778943 0.27745005
#> [15,] 0.011858376 0.14111499 0.11722804 0.08449371 0.14361719 0.16397946
#> [16,] 0.017599840 0.18549684 0.17174644 0.18722415 0.24171948 0.32294115
#> [17,] 0.014911807 0.11365663 0.12942003 0.12210860 0.17525079 0.20025014
#> [18,] 0.007106093 0.09045565 0.09888123 0.12171460 0.15197234 0.15848522
#> [19,] 0.020350257 0.12309913 0.08362986 0.03975478 0.14256782 0.31431943
#> [20,] 0.009489133 0.07856870 0.15206492 0.08457967 0.02974453 0.18231307
#> [21,] 0.030139216 0.12834406 0.18246501 0.19510124 0.25004449 0.36008886
#> [22,] 0.035609975 0.06626557 0.06044243 0.21994086 0.18420355 0.11658798
#> [23,] 0.022898465 0.61680504 0.64162262 0.73399427 0.57874882 0.53808408
#> [24,] 0.029300000 0.08600000 0.04140000 0.04720000 0.08350000 0.09380000
#> [25,] 0.015060352 0.13427682 0.17840612 0.19067899 0.22517992 0.27748928
#>             V15       V16       V17       V18       V19            V2       V20
#>  [1,] 0.2029632 0.1476377 0.2294252 0.2727413 0.3596454  0.0308852071 0.4101539
#>  [2,] 0.3915198 0.5519047 0.6695870 0.7139783 0.7932852  0.0195501376 0.8734487
#>  [3,] 0.3850622 0.4595840 0.5275965 0.6111054 0.6766844  0.0250504839 0.7609949
#>  [4,] 0.1622673 0.2415885 0.2540688 0.3771664 0.5638746  0.0292115145 0.6831675
#>  [5,] 0.2273126 0.2610339 0.3136200 0.2487162 0.2517185  0.0606560659 0.3272673
#>  [6,] 0.3011449 0.3058193 0.2989492 0.2996771 0.4531311  0.0357623233 0.6955028
#>  [7,] 0.1391526 0.1878221 0.2273876 0.2719801 0.2329829  0.0348690948 0.2262930
#>  [8,] 0.5684203 0.7081573 0.9015391 0.9343299 0.8990312  0.1009338498 0.8275468
#>  [9,] 0.3277363 0.3648280 0.1983653 0.2305102 0.3646779  0.0717507034 0.7384751
#> [10,] 0.2363687 0.3968867 0.4835115 0.6716812 0.8164839 -0.0006072355 0.8711463
#> [11,] 0.2422255 0.2426125 0.2178292 0.2124519 0.3293368  0.0304379714 0.3765593
#> [12,] 0.3915277 0.4385213 0.5606726 0.6530494 0.7243314  0.0566411843 0.7983768
#> [13,] 0.5357528 0.5287079 0.6889040 0.7853393 0.8369237  0.0448591194 0.8786924
#> [14,] 0.2243822 0.2412927 0.1596527 0.3074775 0.7266768  0.0243072732 0.9477311
#> [15,] 0.2010391 0.2021889 0.1909196 0.2640582 0.3206330  0.0140708388 0.3613970
#> [16,] 0.3529549 0.4588856 0.5548356 0.5262227 0.3871773  0.0222403965 0.4288953
#> [17,] 0.2311869 0.2522713 0.3081858 0.3614586 0.4105583  0.0121238730 0.4701770
#> [18,] 0.1951962 0.1811082 0.1565042 0.1751311 0.1316899  0.0104685260 0.1517838
#> [19,] 0.5448434 0.7696210 0.8218706 0.8009534 0.8534954  0.0142464822 0.8704446
#> [20,] 0.2873096 0.1984476 0.1294465 0.1707679 0.1758400  0.0214442820 0.2230608
#> [21,] 0.4883214 0.6453583 0.8149145 0.8734367 0.8327099  0.0333949675 0.7501883
#> [22,] 0.1583974 0.2412662 0.2560883 0.2255493 0.2396104  0.0463146423 0.2219420
#> [23,] 0.6601807 0.7078068 0.6971161 0.7773041 0.9154355  0.0574183865 0.8594551
#> [24,] 0.1466000 0.0809000 0.1179000 0.2179000 0.3326000  0.0644000000 0.3258000
#> [25,] 0.1938252 0.2053273 0.2535937 0.3349797 0.4076701  0.0080762666 0.4577412
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.4474900 0.43140346 0.5280004 0.6931389 0.76775192 0.7996716 0.9252526
#>  [2,] 0.9094205 0.94074998 0.8492793 0.5612392 0.33612180 0.3407841 0.3368412
#>  [3,] 0.8161948 0.86616069 0.8895119 0.9217479 0.96076153 0.9599040 0.9275816
#>  [4,] 0.7895156 0.76913862 0.8249942 0.9334358 0.90064322 0.9109615 0.9547390
#>  [5,] 0.3661071 0.48356974 0.5763589 0.6032759 0.60477314 0.6387909 0.7076620
#>  [6,] 0.7982738 0.79282200 0.6909810 0.7169891 0.73955199 0.8620238 0.9698234
#>  [7,] 0.3998965 0.65874365 0.7951528 0.8458364 0.91099623 0.9501035 0.9710610
#>  [8,] 0.7913930 0.51902662 0.2513712 0.2855898 0.26346521 0.1508315 0.1507941
#>  [9,] 0.7922929 0.53010499 0.3852660 0.3321939 0.42325298 0.5903026 0.5816108
#> [10,] 0.8984968 0.95495632 1.0108633 0.9926290 0.96376204 0.8760036 0.7229949
#> [11,] 0.3780112 0.32328932 0.3700047 0.4302241 0.47736054 0.6050130 0.7148404
#> [12,] 0.7805535 0.84545657 0.9755872 1.0476555 0.88912361 0.6324749 0.4476763
#> [13,] 0.9538515 0.99138722 0.9153947 0.9267202 0.96489176 0.9136426 0.8240550
#> [14,] 0.8210013 0.72123993 0.4962791 0.4766691 0.64786833 0.8238356 0.9582662
#> [15,] 0.4069747 0.50797836 0.6176731 0.6978107 0.68526564 0.7351151 0.7869975
#> [16,] 0.5336090 0.61357843 0.7013945 0.7452947 0.72591223 0.6935859 0.6899815
#> [17,] 0.5232834 0.59714269 0.6810569 0.7432021 0.84417065 0.8724742 0.8858160
#> [18,] 0.2207915 0.18963077 0.3219381 0.5741906 0.84923391 1.0000000 0.8501216
#> [19,] 0.9435497 0.93618509 0.9320539 0.9712138 0.94417012 0.9422846 0.9394552
#> [20,] 0.2497669 0.46155135 0.6737195 0.6996384 0.76742185 0.8224017 0.7108318
#> [21,] 0.8071393 0.85840555 0.7905578 0.7794972 0.83063200 0.8393230 0.6608448
#> [22,] 0.2689120 0.09061541 0.1872616 0.2118757 0.06209892 0.1713329 0.4630367
#> [23,] 0.7851130 0.69275886 0.4498158 0.3860599 0.55194001 0.2314954 0.5245769
#> [24,] 0.2111000 0.23020000 0.3361000 0.4259000 0.46090000 0.2606000 0.0874000
#> [25,] 0.5481404 0.59043259 0.5855455 0.7076149 0.94166399 0.9740844 1.0059188
#>             V28       V29          V3       V30       V31       V32       V33
#>  [1,] 0.9895780 0.9252873 0.024877311 0.7621560 0.6595575 0.4670451 0.3693678
#>  [2,] 0.3786386 0.5014906 0.008800693 0.5551352 0.3910520 0.4324413 0.4690453
#>  [3,] 0.8187599 0.6920572 0.024141988 0.5503290 0.3855126 0.3176458 0.2492846
#>  [4,] 0.9818050 0.8508144 0.039538505 0.6133995 0.4017290 0.2698487 0.2017575
#>  [5,] 0.8352552 0.8528301 0.065539143 0.8567644 0.8968380 0.8780490 0.7533884
#>  [6,] 0.9423112 0.8046901 0.051646228 0.5816428 0.4929745 0.5528859 0.4640506
#>  [7,] 0.9599326 0.8059508 0.024877127 0.6641678 0.6787028 0.5977507 0.4160422
#>  [8,] 0.1751676 0.1545503 0.122432704 0.1303227 0.2402314 0.2957489 0.3343790
#>  [9,] 0.7394645 0.6579396 0.036213031 0.6432078 0.3396881 0.2990941 0.4270968
#> [10,] 0.5835354 0.5014372 0.026962712 0.4897948 0.4075530 0.2537862 0.1785826
#> [11,] 0.8132980 0.8747298 0.036202782 0.8312192 0.6739227 0.5538903 0.5402629
#> [12,] 0.3563762 0.3168709 0.074213365 0.2672727 0.3269436 0.3546548 0.4541656
#> [13,] 0.7023274 0.6430881 0.040100335 0.4720131 0.2685322 0.1620690 0.1128426
#> [14,] 0.8018735 0.5426077 0.018572653 0.3667899 0.3542211 0.2768595 0.3271935
#> [15,] 0.8792296 0.9460759 0.025254197 0.8834922 0.8091635 0.7171716 0.6811881
#> [16,] 0.5968293 0.4540386 0.041285595 0.4410135 0.5302469 0.5220714 0.4698894
#> [17,] 0.8856294 0.7478116 0.016584279 0.5945533 0.5328314 0.4344044 0.3681198
#> [18,] 0.6646984 0.5558741 0.012725571 0.4063880 0.3306238 0.2054195 0.1861565
#> [19,] 0.7481007 0.4570439 0.030522098 0.3488838 0.4021784 0.4082098 0.3459933
#> [20,] 0.7580221 0.7489600 0.020898653 0.6384143 0.4843598 0.3374198 0.5357599
#> [21,] 0.4492198 0.4347468 0.044410615 0.5314844 0.4536109 0.2692968 0.2768131
#> [22,] 0.7367535 0.8071703 0.039669221 0.7276802 0.7240530 0.5638147 0.6313159
#> [23,] 0.8738471 0.5982100 0.114570922 0.8526381 0.8773747 0.5126538 0.1618512
#> [24,] 0.2862000 0.5606000 0.039000000 0.8344000 0.8096000 0.7250000 0.8048000
#> [25,] 0.9850157 0.8791368 0.010885394 0.7187012 0.6015102 0.4492306 0.3193873
#>              V34        V35        V36        V37        V38       V39
#>  [1,] 0.32450079 0.20897118 0.13703390 0.20867754 0.15303966 0.1878981
#>  [2,] 0.32813539 0.21960763 0.22811434 0.21497217 0.25550716 0.2140913
#>  [3,] 0.12510870 0.11804501 0.17532415 0.24764930 0.29002158 0.2777068
#>  [4,] 0.15117627 0.11903670 0.03059587 0.06338429 0.09529729 0.1020414
#>  [5,] 0.64937723 0.63518960 0.47927305 0.35713800 0.37545646 0.3221410
#>  [6,] 0.28938115 0.07212021 0.27146930 0.28116314 0.37789356 0.5570063
#>  [7,] 0.16469216 0.09990515 0.09099104 0.13353407 0.15455138 0.1653486
#>  [8,] 0.43292527 0.50258677 0.24161092 0.13836950 0.24144852 0.2743378
#>  [9,] 0.60597529 0.78340442 0.82128336 0.82300504 0.78958748 0.6674277
#> [10,] 0.26089978 0.35039640 0.35444100 0.26188843 0.22252735 0.3343010
#> [11,] 0.43155625 0.26978180 0.20942916 0.25198534 0.28285083 0.3111641
#> [12,] 0.48900273 0.47569537 0.32072699 0.27231262 0.37415185 0.3731661
#> [13,] 0.08780344 0.12780976 0.20460060 0.22288925 0.31457029 0.3694399
#> [14,] 0.36311627 0.35140181 0.31384135 0.41308037 0.48727506 0.4448877
#> [15,] 0.63675548 0.57461327 0.53351204 0.40156152 0.28648755 0.2211095
#> [16,] 0.41955455 0.51053258 0.61587384 0.62017640 0.43190523 0.3415132
#> [17,] 0.28340644 0.27806487 0.34703556 0.31499267 0.26403436 0.2442937
#> [18,] 0.22007043 0.31378452 0.22828035 0.08462279 0.06769042 0.1041728
#> [19,] 0.27977825 0.22983767 0.12580973 0.23541240 0.20048785 0.1251707
#> [20,] 0.86826211 0.85411004 0.67144989 0.35086804 0.31983013 0.2715166
#> [21,] 0.33026167 0.24202877 0.15562799 0.15119180 0.11163488 0.1269036
#> [22,] 0.82450803 0.86167880 0.84961675 0.71698946 0.66358021 0.6443137
#> [23,] 0.23754699 0.39987040 0.27846924 0.60572227 0.66689595 0.5235810
#> [24,] 0.94350000 1.00000000 0.89600000 0.55160000 0.30370000 0.2338000
#> [25,] 0.18672281 0.08232700 0.24327295 0.24935288 0.32200539 0.3525085
#>                V4        V40        V41        V42        V43        V44
#>  [1,] 0.057814651 0.29548830 0.27208522 0.29122722 0.18842076 0.14029615
#>  [2,] 0.038131834 0.14376145 0.14062614 0.19795581 0.18877083 0.18405234
#>  [3,] 0.050328178 0.29075276 0.26963963 0.26792021 0.25509665 0.21385264
#>  [4,] 0.045119170 0.11784477 0.07352702 0.10620961 0.19385178 0.21581282
#>  [5,] 0.074889121 0.33477989 0.56412956 0.57652953 0.39705263 0.42607919
#>  [6,] 0.041385187 0.44749934 0.23718525 0.42356371 0.44841690 0.48372629
#>  [7,] 0.036914407 0.08792966 0.12549603 0.09663687 0.14813184 0.14443224
#>  [8,] 0.144493737 0.13568886 0.06398669 0.05202120 0.10653329 0.07360227
#>  [9,] 0.069789655 0.46751351 0.30109920 0.27356942 0.31378231 0.33775093
#> [10,] 0.076407919 0.35978456 0.37201751 0.44821093 0.33197353 0.19578253
#> [11,] 0.056967045 0.38961908 0.42820427 0.41223202 0.37575551 0.28605625
#> [12,] 0.039555270 0.40342927 0.33106890 0.35473084 0.33792561 0.27008862
#> [13,] 0.023600783 0.26664129 0.20679173 0.20127319 0.07660453 0.14400693
#> [14,] 0.009312742 0.33704644 0.20058312 0.14678418 0.10909185 0.06920715
#> [15,] 0.024091148 0.17478741 0.17615846 0.19180434 0.22590721 0.18617720
#> [16,] 0.035707092 0.45086764 0.44331239 0.35523565 0.24653244 0.18544764
#> [17,] 0.020402155 0.16643091 0.15836120 0.19656423 0.17535022 0.12079349
#> [18,] 0.033869384 0.18369372 0.25453721 0.23916885 0.19058660 0.09468313
#> [19,] 0.030522021 0.12538387 0.08116624 0.04882915 0.08042520 0.11993020
#> [20,] 0.049395612 0.25709745 0.26703725 0.34641079 0.42229989 0.23454603
#> [21,] 0.042477062 0.22813252 0.22303968 0.16078816 0.08570182 0.11799661
#> [22,] 0.022080662 0.63232313 0.55163492 0.47911452 0.31830222 0.15843702
#> [23,] 0.104692072 0.46367743 0.46528152 0.24530627 0.19821192 0.21282064
#> [24,] 0.017300000 0.23820000 0.33180000 0.38210000 0.15750000 0.22280000
#> [25,] 0.025191526 0.22857458 0.08912268 0.21827003 0.18468237 0.15965119
#>              V45          V46        V47        V48         V49         V5
#>  [1,] 0.19785472  0.139588450 0.16014297 0.12424854 0.059212552 0.09820315
#>  [2,] 0.13463589  0.069166122 0.07640054 0.07093358 0.044803061 0.05693646
#>  [3,] 0.16260316  0.105599815 0.09332972 0.07032009 0.050332401 0.08584523
#>  [4,] 0.18809555  0.208618573 0.17549418 0.10929530 0.052019462 0.06270266
#>  [5,] 0.53663987  0.463600522 0.27579388 0.15893750 0.082022039 0.07685333
#>  [6,] 0.48700681  0.219010768 0.14033977 0.16337762 0.089032997 0.08463298
#>  [7,] 0.09645227  0.090389282 0.11754371 0.10687234 0.065733188 0.04784602
#>  [8,] 0.04167437  0.038125365 0.08941733 0.09640269 0.060303094 0.13266881
#>  [9,] 0.44521314  0.350537790 0.16484269 0.10874078 0.068358623 0.06071501
#> [10,] 0.07663540  0.168265617 0.20098634 0.11798764 0.059306948 0.10576423
#> [11,] 0.32510135  0.301477503 0.22931493 0.17317580 0.100393923 0.07963927
#> [12,] 0.20283997  0.139279525 0.10216176 0.06095451 0.042708648 0.08328228
#> [13,] 0.15049391  0.124957750 0.12456538 0.09894318 0.062087463 0.11315247
#> [14,] 0.10343546  0.098634125 0.07483789 0.05015746 0.020586956 0.03673774
#> [15,] 0.10551926  0.080661253 0.05341204 0.04886149 0.024937962 0.02909737
#> [16,] 0.13172381  0.125766590 0.08855459 0.06827006 0.041412860 0.04848573
#> [17,] 0.10605629  0.098084012 0.07056681 0.05946317 0.023506720 0.04294890
#> [18,] 0.09132748  0.090753904 0.03734035 0.04416226 0.034009044 0.04963026
#> [19,] 0.10606590  0.047650019 0.04844750 0.04334081 0.024385860 0.07299259
#> [20,] 0.06663620  0.014048767 0.07413680 0.07017757 0.034180334 0.06508796
#> [21,] 0.13446018  0.098769936 0.12902048 0.09522157 0.045136020 0.08536512
#> [22,] 0.08113481  0.026458736 0.07365481 0.07465640 0.038951548 0.03517321
#> [23,] 0.19495499 -0.005304401 0.13842669 0.06987915 0.008954369 0.09517396
#> [24,] 0.15820000  0.143300000 0.16340000 0.11330000 0.056700000 0.04760000
#> [25,] 0.11682662  0.081039046 0.11596594 0.09955368 0.045630730 0.01697514
#>               V50          V51         V52         V53         V54         V55
#>  [1,] 0.007023245 0.0217598416 0.014395630 0.009110336 0.008008328 0.004265259
#>  [2,] 0.016052069 0.0151828970 0.013395350 0.003409414 0.005460459 0.003008901
#>  [3,] 0.021999535 0.0168538077 0.010784524 0.014626648 0.008618978 0.006462106
#>  [4,] 0.015151080 0.0145158661 0.011976711 0.005189722 0.005625760 0.005080938
#>  [5,] 0.035090619 0.0289733634 0.029458653 0.017139198 0.014131224 0.011965437
#>  [6,] 0.018151538 0.0288740852 0.017711203 0.007700310 0.010669172 0.011112987
#>  [7,] 0.024574087 0.0157600273 0.011664044 0.007664079 0.012276710 0.004309938
#>  [8,] 0.022269137 0.0238008651 0.028653513 0.016037149 0.012177964 0.018367245
#>  [9,] 0.029847030 0.0151838373 0.018490556 0.012980530 0.014994263 0.014160725
#> [10,] 0.032897058 0.0256916663 0.022149616 0.027592535 0.017723792 0.003146134
#> [11,] 0.027005198 0.0130434476 0.013617215 0.008446640 0.005677075 0.007592680
#> [12,] 0.010840059 0.0001281006 0.008973752 0.017328452 0.012391869 0.005747809
#> [13,] 0.027792779 0.0211176082 0.016848413 0.014078895 0.026700278 0.016651479
#> [14,] 0.024665252 0.0148776368 0.014066292 0.020629284 0.021257621 0.011342893
#> [15,] 0.014869198 0.0106423362 0.010983203 0.008345075 0.005183869 0.005079435
#> [16,] 0.014250302 0.0160712594 0.007547104 0.007036147 0.008606398 0.007437811
#> [17,] 0.017010185 0.0063216858 0.006053095 0.006051482 0.006646382 0.005706276
#> [18,] 0.007077742 0.0079584349 0.012309044 0.005285739 0.003831131 0.004320697
#> [19,] 0.006296598 0.0166302713 0.011407205 0.008748135 0.007065028 0.004057372
#> [20,] 0.014984323 0.0291070269 0.010998776 0.010877215 0.010069061 0.002805546
#> [21,] 0.019359900 0.0128800364 0.006981624 0.007916859 0.008289334 0.006814438
#> [22,] 0.015343933 0.0074769420 0.004456589 0.020196783 0.014919838 0.005907453
#> [23,] 0.008330194 0.0022995348 0.022787308 0.015199349 0.008255671 0.018031265
#> [24,] 0.013300000 0.0170000000 0.003500000 0.005200000 0.008300000 0.007800000
#> [25,] 0.014362185 0.0032294977 0.004588498 0.003247581 0.003907446 0.002449793
#>               V56         V57         V58         V59         V6         V60
#>  [1,] 0.008578691 0.005089758 0.005332382 0.002311651 0.14486416 0.003227720
#>  [2,] 0.005523798 0.003337337 0.006652264 0.002631063 0.10024787 0.001848077
#>  [3,] 0.009786058 0.007197564 0.006748223 0.006655875 0.13261538 0.004997223
#>  [4,] 0.004395592 0.003947010 0.001698832 0.001332022 0.10543023 0.004596271
#>  [5,] 0.008495081 0.009594258 0.008266730 0.009528199 0.12506416 0.006411074
#>  [6,] 0.003848122 0.006004692 0.003264438 0.004948696 0.12104955 0.003611187
#>  [7,] 0.008531680 0.006927705 0.009192747 0.011580809 0.12555873 0.010526360
#>  [8,] 0.013894662 0.006243500 0.014318981 0.019614500 0.06591929 0.012257063
#>  [9,] 0.009869088 0.008049149 0.009386698 0.008290696 0.07381788 0.008622461
#> [10,] 0.006848146 0.012365485 0.007670393 0.007679684 0.07664092 0.006689136
#> [11,] 0.008195239 0.006507478 0.008876289 0.009996850 0.08318059 0.004869768
#> [12,] 0.014053315 0.007827107 0.007758585 0.010291349 0.12030435 0.006329108
#> [13,] 0.008084324 0.007637930 0.005989203 0.007067153 0.17008562 0.006061164
#> [14,] 0.009856729 0.005396314 0.009307618 0.001794858 0.02477399 0.003087777
#> [15,] 0.004428961 0.004662832 0.004590096 0.009298139 0.06617591 0.006212420
#> [16,] 0.007800051 0.009281713 0.005393265 0.008363606 0.11353905 0.005965758
#> [17,] 0.005168459 0.007686381 0.007803993 0.004339689 0.06503042 0.004994783
#> [18,] 0.000731131 0.002364870 0.004413043 0.002107826 0.11167494 0.001619478
#> [19,] 0.007296579 0.005548747 0.004702224 0.006224372 0.12389390 0.004526170
#> [20,] 0.009663175 0.009668560 0.004252187 0.006989222 0.07888927 0.010589638
#> [21,] 0.008130599 0.006865990 0.004067764 0.005422013 0.11537215 0.006847320
#> [22,] 0.009336112 0.008619520 0.009880321 0.010674608 0.06153405 0.010003780
#> [23,] 0.024368735 0.032694259 0.016969015 0.009937704 0.24084347 0.007656183
#> [24,] 0.007500000 0.010500000 0.016000000 0.009500000 0.08160000 0.001100000
#> [25,] 0.004260560 0.003907971 0.001260167 0.002559082 0.03373945 0.002202177
#>               V7         V8         V9
#>  [1,] 0.12630463 0.10318172 0.15419372
#>  [2,] 0.11029782 0.10680502 0.18935906
#>  [3,] 0.14456044 0.13329421 0.18957945
#>  [4,] 0.14976978 0.15889826 0.19159448
#>  [5,] 0.10510683 0.12618464 0.25635824
#>  [6,] 0.12352643 0.18133293 0.27689373
#>  [7,] 0.15118854 0.15506929 0.20457500
#>  [8,] 0.09696932 0.07672695 0.14746837
#>  [9,] 0.11835584 0.23615767 0.34313291
#> [10,] 0.13856316 0.11083750 0.03058289
#> [11,] 0.09561114 0.14765356 0.21665931
#> [12,] 0.13121789 0.04678178 0.07003473
#> [13,] 0.17232801 0.15525627 0.18601961
#> [14,] 0.03307062 0.02123019 0.07885535
#> [15,] 0.05748439 0.07896694 0.11635390
#> [16,] 0.12120095 0.15001643 0.16261054
#> [17,] 0.07241040 0.05123866 0.09323150
#> [18,] 0.11773582 0.10030156 0.12218591
#> [19,] 0.12164113 0.07750726 0.12582092
#> [20,] 0.09465941 0.08614454 0.07147124
#> [21,] 0.13394887 0.09778103 0.11614885
#> [22,] 0.03028647 0.06412612 0.08202713
#> [23,] 0.25347983 0.38393443 0.55938731
#> [24,] 0.09930000 0.03150000 0.07360000
#> [25,] 0.08614019 0.08427177 0.09313750
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768