Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13        V14
#>  [1,] 0.03593584 0.24817282 0.29754527 0.28905898 0.19483128 0.05600527
#>  [2,] 0.04897866 0.26200895 0.38139959 0.48824458 0.56607854 0.61042048
#>  [3,] 0.02011602 0.24046877 0.28935548 0.31126722 0.36889852 0.49458275
#>  [4,] 0.02770670 0.21415467 0.25798417 0.29525924 0.32963955 0.37713636
#>  [5,] 0.04361320 0.28491790 0.21870022 0.11706957 0.15359363 0.31409160
#>  [6,] 0.04049204 0.35660561 0.37795982 0.41212270 0.49675982 0.48609600
#>  [7,] 0.01643622 0.18465823 0.28459557 0.31699952 0.31923506 0.27450377
#>  [8,] 0.04387467 0.04466590 0.12988559 0.16195988 0.18453495 0.29112868
#>  [9,] 0.02296825 0.24551805 0.28713656 0.29970141 0.28365474 0.25737527
#> [10,] 0.01663160 0.20317853 0.26194896 0.29996621 0.25063023 0.16430060
#> [11,] 0.01735015 0.24948707 0.28181301 0.34425842 0.30706436 0.23271520
#> [12,] 0.04323617 0.28396485 0.30677900 0.31989801 0.27352944 0.16589504
#> [13,] 0.01529726 0.29575696 0.38455625 0.38279457 0.40456707 0.34922401
#> [14,] 0.02556255 0.49166700 0.55935897 0.51609772 0.46895280 0.35040388
#> [15,] 0.01602256 0.03631987 0.06498999 0.09724415 0.14315462 0.13685311
#> [16,] 0.01909150 0.07885791 0.15488103 0.18302947 0.26074421 0.24148364
#> [17,] 0.01687280 0.08727950 0.07715334 0.05960907 0.06854909 0.09460908
#> [18,] 0.01317465 0.11098942 0.12979766 0.10403329 0.19904859 0.35825661
#> [19,] 0.04227667 0.30732702 0.27260127 0.13261898 0.21117223 0.45039114
#> [20,] 0.02544646 0.06443927 0.05764557 0.14267285 0.15714598 0.07221724
#> [21,] 0.01498630 0.10978175 0.08949913 0.05220461 0.07108516 0.10114007
#> [22,] 0.02183765 0.13990603 0.15376090 0.16560431 0.20278542 0.18724999
#> [23,] 0.02374640 0.14009828 0.15139549 0.15612466 0.23797851 0.37725746
#> [24,] 0.01083316 0.04643353 0.13506131 0.17464202 0.23795275 0.32009536
#> [25,] 0.01750862 0.19597062 0.15789274 0.08866750 0.09570277 0.12794341
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.11574017 0.1770920 0.2279779 0.3250332 0.3824371 0.049884387 0.4147684
#>  [2,] 0.68657866 0.7703142 0.9023446 0.9277498 0.9235948 0.058823741 0.8444691
#>  [3,] 0.52802267 0.5595695 0.6104724 0.7366653 0.8010188 0.023802534 0.8491575
#>  [4,] 0.45238098 0.5339364 0.6093534 0.6663851 0.7180236 0.033529699 0.7652506
#>  [5,] 0.49651000 0.6042887 0.6212802 0.5464229 0.5875644 0.039460051 0.8404709
#>  [6,] 0.30932775 0.3340023 0.4018027 0.3961355 0.5561657 0.026109741 0.7096319
#>  [7,] 0.25712906 0.1436119 0.1867792 0.2465075 0.2923197 0.022296053 0.3516636
#>  [8,] 0.37438871 0.4624857 0.5260996 0.6325264 0.7213474 0.041320616 0.7924843
#>  [9,] 0.18628501 0.1945984 0.2473846 0.2397652 0.3529699 0.037422086 0.3678508
#> [10,] 0.07644841 0.2763504 0.4431207 0.6279924 0.7875027 0.003055155 0.7713372
#> [11,] 0.17196463 0.2214782 0.2522415 0.3704266 0.5834609 0.020382619 0.6915322
#> [12,] 0.10742325 0.2049932 0.2079632 0.2170731 0.2586638 0.023565590 0.3159886
#> [13,] 0.11453810 0.1137862 0.3541900 0.5513095 0.6780870 0.020673962 0.7530567
#> [14,] 0.25474179 0.2861790 0.2813390 0.3162658 0.4863171 0.039214427 0.6890755
#> [15,] 0.16961098 0.2593291 0.1739140 0.2601017 0.4590261 0.040140040 0.1886195
#> [16,] 0.32458276 0.4114671 0.3277249 0.3722201 0.5294560 0.026630149 0.5561852
#> [17,] 0.02936311 0.1225849 0.1565041 0.3620203 0.3510446 0.017672408 0.3081004
#> [18,] 0.49244216 0.6692626 0.7838304 0.8098129 0.8499387 0.014682525 0.8359888
#> [19,] 0.58970147 0.7023347 0.7906059 0.7019243 0.4550049 0.063230997 0.3934942
#> [20,] 0.14268120 0.1399118 0.1890945 0.2594349 0.2695809 0.047679732 0.2644697
#> [21,] 0.09248953 0.1108368 0.1802498 0.2635088 0.2813182 0.010270682 0.3646496
#> [22,] 0.27149705 0.5943274 0.6915279 0.5421023 0.3475335 0.008530244 0.3349699
#> [23,] 0.58628950 0.7831652 0.8721361 0.8451198 0.8558477 0.024382915 0.8533609
#> [24,] 0.41942869 0.4705671 0.5479638 0.6454868 0.7053005 0.008473568 0.7850584
#> [25,] 0.18611370 0.2338889 0.2392330 0.1987445 0.3497342 0.002738916 0.5654961
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.5371501 0.7457856 0.8482109 0.8968562 0.9284801 0.9520815 0.8657671
#>  [2,] 0.7687648 0.6740649 0.5425446 0.4079126 0.3279354 0.2935120 0.2858129
#>  [3,] 0.9151763 0.9607492 0.9031359 0.8743696 0.9095046 0.8914444 0.7814414
#>  [4,] 0.8145424 0.8706231 0.9237614 0.9704713 0.9405836 0.9342683 0.9207424
#>  [5,] 0.9742426 0.8716888 0.6766680 0.4594077 0.2948873 0.3572644 0.3828871
#>  [6,] 0.6607757 0.7937810 0.8623816 0.8230234 0.8859735 0.9644336 0.9533172
#>  [7,] 0.4405434 0.4727354 0.5448896 0.6478065 0.7370361 0.7834137 0.9604332
#>  [8,] 0.8838471 0.9800816 1.0070170 1.0179441 0.8978734 0.7179829 0.5956535
#>  [9,] 0.4029939 0.4091410 0.4541383 0.5311693 0.5937564 0.6938742 0.8223738
#> [10,] 0.7933549 0.8481584 0.9456900 0.9675146 0.9954823 0.8339003 0.5160254
#> [11,] 0.8025716 0.7953106 0.8307088 0.9542839 0.9625666 0.9661170 0.9782006
#> [12,] 0.4474582 0.6343215 0.7905260 0.8216082 0.8850646 0.9612803 0.9945564
#> [13,] 0.7987869 0.8344468 0.8689836 0.7755114 0.7049986 0.5646942 0.2740903
#> [14,] 0.8111890 0.7902907 0.7176796 0.7527283 0.7109822 0.7962285 0.8720813
#> [15,] 0.3925250 0.4599716 0.3841484 0.5301623 0.4424004 0.5196875 0.3599606
#> [16,] 0.4889160 0.4114441 0.5600363 0.7449675 0.7847992 0.8497041 0.8386633
#> [17,] 0.2138485 0.1794420 0.1802133 0.2338285 0.0685357 0.1953122 0.4218515
#> [18,] 0.8407305 0.8644879 0.8609860 0.8263269 0.8183742 0.8624412 0.9529636
#> [19,] 0.5361284 0.6951365 0.7893061 0.8064315 0.7134469 0.6121509 0.6513747
#> [20,] 0.2032363 0.1729378 0.2483379 0.2760250 0.2385120 0.2143316 0.2564519
#> [21,] 0.4631143 0.5920445 0.7183083 0.7651892 0.7759227 0.8332972 0.8560077
#> [22,] 0.4146410 0.4885624 0.5460949 0.5749625 0.7616248 0.9440626 0.9398320
#> [23,] 0.9402319 0.9347858 0.8680731 0.8670204 0.8818830 0.9529534 0.8532656
#> [24,] 0.9319679 0.9832494 0.9361712 0.9197760 0.8703050 0.7950409 0.6243007
#> [25,] 0.6204953 0.8097444 1.0383883 0.9059194 0.7539777 0.8216407 0.8514513
#>             V28       V29          V3       V30       V31       V32       V33
#>  [1,] 0.7113000 0.7027829 0.050402121 0.7805667 0.7572528 0.5201487 0.2858917
#>  [2,] 0.2751735 0.1989890 0.068102643 0.2315385 0.2290797 0.2664012 0.3058108
#>  [3,] 0.6298771 0.6149798 0.022257674 0.4942349 0.3069630 0.1850153 0.1532030
#>  [4,] 0.8803982 0.7467756 0.035204985 0.5785234 0.4375938 0.3318383 0.2505734
#>  [5,] 0.2977172 0.2982467 0.020464731 0.5047736 0.2500957 0.3991200 0.4953527
#>  [6,] 0.7494569 0.4369887 0.027857123 0.4718497 0.4910030 0.5747706 0.4979725
#>  [7,] 1.1183191 0.9805235 0.031349054 0.7064051 0.7005212 0.5863414 0.3734739
#>  [8,] 0.4930764 0.3979287 0.072523473 0.3990285 0.3282295 0.2127204 0.2415354
#>  [9,] 0.9019119 0.8942767 0.042154587 0.8425909 0.7678015 0.6416053 0.5354868
#> [10,] 0.3674412 0.2886352 0.013739105 0.4948825 0.6811929 0.7901887 0.8645357
#> [11,] 0.9826952 0.8815312 0.034671117 0.6506180 0.4065572 0.2880670 0.2187629
#> [12,] 1.0167526 0.9369401 0.026631762 0.8214278 0.6960158 0.4701429 0.4560278
#> [13,] 0.4733998 0.6659117 0.020830824 0.7678550 0.8346051 0.8639327 1.0166237
#> [14,] 0.8913637 0.7698659 0.057499683 0.5338498 0.4328705 0.4334517 0.4217182
#> [15,] 0.3276701 0.5854144 0.041998803 0.7467897 0.7474279 0.4781494 0.4089003
#> [16,] 0.8172916 0.7156248 0.031554798 0.6174760 0.5097683 0.3301509 0.3082474
#> [17,] 0.6800645 0.8393912 0.016043265 0.7593020 0.5299664 0.1992396 0.4821875
#> [18,] 0.9076214 0.7142412 0.009983381 0.5256563 0.3178054 0.2651391 0.1698384
#> [19,] 0.6085695 0.4138526 0.080759414 0.4247747 0.4519407 0.3544014 0.2360558
#> [20,] 0.4497845 0.6509985 0.033114498 0.7983323 0.8154852 0.7181100 0.7561909
#> [21,] 0.8604899 0.8873538 0.016982266 0.8166207 0.7397460 0.6369937 0.5924312
#> [22,] 0.8525041 0.6540278 0.016789458 0.6880128 0.8090297 0.7009205 0.6076111
#> [23,] 0.6314850 0.4402674 0.034894559 0.4485465 0.4851276 0.3699445 0.3207314
#> [24,] 0.3914034 0.4553735 0.006162668 0.5869198 0.4849296 0.3103016 0.3496293
#> [25,] 0.8548893 0.7623233 0.009397816 0.6565028 0.6112784 0.4813402 0.4098236
#>              V34       V35        V36        V37        V38        V39
#>  [1,] 0.19745255 0.1560817 0.19417364 0.27115473 0.16016068 0.11634801
#>  [2,] 0.29865358 0.3332435 0.16194673 0.22640149 0.33046362 0.27634359
#>  [3,] 0.06730181 0.1024272 0.16880375 0.20516020 0.27576024 0.37938188
#>  [4,] 0.15339248 0.1423702 0.12397884 0.17820725 0.23236544 0.19713381
#>  [5,] 0.32496469 0.2405551 0.38674843 0.42636621 0.45958195 0.35180103
#>  [6,] 0.37878584 0.1367254 0.22309027 0.31179219 0.33598057 0.36705659
#>  [7,] 0.30262822 0.1692235 0.08374656 0.21207396 0.16981614 0.23601465
#>  [8,] 0.28630919 0.3570841 0.39421175 0.35831068 0.40050168 0.36758912
#>  [9,] 0.45000011 0.3237444 0.26429054 0.29119337 0.25289988 0.24761477
#> [10,] 0.82502296 0.6375533 0.43618707 0.13800070 0.15101339 0.30941681
#> [11,] 0.17753400 0.1551427 0.07328999 0.06467990 0.09133705 0.09120555
#> [12,] 0.28871341 0.1250379 0.01919014 0.08958836 0.15543567 0.13295908
#> [13,] 0.97293698 0.6837387 0.37863560 0.09533046 0.35535926 0.64352165
#> [14,] 0.35921230 0.1943933 0.31632072 0.28582495 0.36649257 0.55857088
#> [15,] 0.46749166 0.5999059 0.65243189 0.62259389 0.77612607 0.95448703
#> [16,] 0.17181895 0.2473113 0.42434235 0.33574410 0.13152185 0.16638664
#> [17,] 0.49297626 0.2463213 0.35301045 0.49942245 0.58403613 0.63234894
#> [18,] 0.21911864 0.2700111 0.23298984 0.20382029 0.17155380 0.07560166
#> [19,] 0.14783026 0.4196950 0.61776557 0.59469437 0.37183974 0.22916780
#> [20,] 0.92331181 0.9979936 0.92419952 0.68742388 0.52062704 0.44757225
#> [21,] 0.60736739 0.5619103 0.46281176 0.31599477 0.22040043 0.16615419
#> [22,] 0.58623300 0.6691846 0.62948236 0.53585734 0.46303751 0.35425804
#> [23,] 0.30380757 0.1832372 0.08927386 0.22421535 0.16810063 0.10033681
#> [24,] 0.32428925 0.1138355 0.09844764 0.08268181 0.06644191 0.16629328
#> [25,] 0.47748027 0.4530401 0.37952745 0.38149252 0.27521602 0.17445983
#>                V4        V40        V41         V42        V43         V44
#>  [1,] 0.059917100 0.13749545 0.15426381  0.26997092 0.28500800  0.19294254
#>  [2,] 0.104076041 0.18437964 0.16824370  0.23472891 0.20254398  0.08110536
#>  [3,] 0.065170192 0.35801059 0.26867515  0.25397312 0.23087866  0.18089575
#>  [4,] 0.055187366 0.20112207 0.24392921  0.22552330 0.17612852  0.16035250
#>  [5,] 0.037002552 0.23237374 0.12715141  0.20556835 0.19784752  0.15014668
#>  [6,] 0.040175307 0.27951129 0.14339082  0.26770635 0.38179832  0.45791978
#>  [7,] 0.059284162 0.26520417 0.25919634  0.28709334 0.11069722  0.07662796
#>  [8,] 0.103288803 0.32496647 0.39798706  0.46810163 0.31720138  0.21130027
#>  [9,] 0.058233336 0.35512461 0.43041746  0.42940221 0.38701270  0.32124603
#> [10,] 0.039891711 0.46094519 0.43108360  0.30830710 0.19527174  0.08555299
#> [11,] 0.035922153 0.09236113 0.05758513  0.11211734 0.20108205  0.23397856
#> [12,] 0.024507648 0.05785808 0.11144543  0.14206385 0.20760274  0.15434943
#> [13,] 0.039996354 0.81832956 0.76875550  0.53753777 0.23666247  0.15830219
#> [14,] 0.056459330 0.43658847 0.23230263  0.40571619 0.44235165  0.43759709
#> [15,] 0.011223148 0.95159938 0.79787885  0.73010052 0.69951905  0.66100243
#> [16,] 0.040618394 0.13370482 0.25077097  0.26962500 0.20634772  0.10817872
#> [17,] 0.026854749 0.47826307 0.22753569  0.17877498 0.12459119  0.04726391
#> [18,] 0.037013220 0.05791517 0.06382462  0.01745725 0.05342062  0.05227245
#> [19,] 0.071033961 0.34333424 0.35299369  0.29721826 0.22256064  0.18749901
#> [20,] 0.006177088 0.46003004 0.46010986  0.45194978 0.24860701  0.20901939
#> [21,] 0.016827687 0.22018972 0.22793309  0.17006712 0.14723278  0.10211429
#> [22,] 0.032686823 0.28703706 0.23648430  0.18382254 0.15903152  0.10840031
#> [23,] 0.039399259 0.16941859 0.11405970  0.08086818 0.08795786  0.13562741
#> [24,] 0.040916796 0.21635375 0.16283620  0.06843373 0.02842897  0.07859776
#> [25,] 0.036322100 0.07684592 0.06000983 -0.01007476 0.04145306 -0.01698517
#>                V45          V46        V47        V48        V49         V5
#>  [1,]  0.152013872  0.169228559 0.16112186 0.08226166 0.05224203 0.04474380
#>  [2,]  0.100370574  0.084407254 0.05446556 0.05323640 0.03723728 0.13818046
#>  [3,]  0.135364292  0.111923535 0.10661156 0.08411621 0.06673966 0.10412814
#>  [4,]  0.141880354  0.092998824 0.08246702 0.06450425 0.03872805 0.11181639
#>  [5,]  0.160329200  0.068461576 0.05726578 0.08048078 0.04987361 0.05378043
#>  [6,]  0.478572893  0.353943438 0.18906926 0.10951526 0.04447814 0.10708153
#>  [7,]  0.142741256  0.223183019 0.24436875 0.17599131 0.05489905 0.07009630
#>  [8,]  0.178707163  0.191733534 0.16966092 0.09542080 0.06069875 0.12064445
#>  [9,]  0.386842851  0.302903830 0.22840751 0.17779590 0.09731228 0.08481590
#> [10,]  0.044503105  0.078087633 0.10408102 0.09994633 0.07164523 0.05162246
#> [11,]  0.220226027  0.226464657 0.18004458 0.11177005 0.05616151 0.07690040
#> [12,]  0.075316912  0.079148787 0.10801330 0.09958939 0.06301024 0.04533738
#> [13,]  0.285709763  0.285666823 0.20483220 0.15420638 0.12759194 0.08046242
#> [14,]  0.392101685  0.153491497 0.13517676 0.17926343 0.10001607 0.09420204
#> [15,]  0.569507662  0.464978870 0.34073046 0.19860795 0.10808869 0.10006769
#> [16,]  0.104756576  0.112720750 0.05970295 0.05429849 0.02258249 0.04382477
#> [17,]  0.003055037 -0.005780680 0.05816409 0.06060508 0.01495359 0.04331527
#> [18,]  0.049501199  0.001049760 0.03827566 0.03237648 0.02019994 0.06276139
#> [19,]  0.180654702  0.122756068 0.12933478 0.11952892 0.07622118 0.09232497
#> [20,]  0.127906723  0.091734826 0.13088187 0.09197010 0.04455744 0.01877881
#> [21,]  0.070264638  0.053236217 0.03457843 0.03025773 0.01489815 0.03015162
#> [22,]  0.074462902  0.105017397 0.04986855 0.03578044 0.01794967 0.04716187
#> [23,]  0.137075258  0.092498139 0.10187862 0.07724587 0.03698995 0.06453736
#> [24,]  0.067309953  0.034005801 0.06483253 0.04663579 0.02487508 0.07082752
#> [25,] -0.059730946  0.002108293 0.03036628 0.02596385 0.01732542 0.05935686
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.018704884 0.022303001 0.020040456 0.013391740 0.008353419 0.004835969
#>  [2,] 0.025833428 0.019768629 0.022307792 0.017562615 0.017404965 0.018894278
#>  [3,] 0.028324940 0.024225441 0.011595784 0.014652829 0.015258384 0.011957530
#>  [4,] 0.016125981 0.015593062 0.015760949 0.014814504 0.013776261 0.006025190
#>  [5,] 0.025115807 0.017103522 0.011243482 0.001793345 0.015975496 0.007692392
#>  [6,] 0.027913976 0.009868757 0.010059370 0.018023052 0.012274081 0.013307992
#>  [7,] 0.013057167 0.022189194 0.014904442 0.007697792 0.007213736 0.006560255
#>  [8,] 0.027074089 0.024201407 0.014940186 0.021624369 0.020188202 0.005789540
#>  [9,] 0.023551177 0.018524656 0.015971172 0.012500701 0.009300062 0.006506260
#> [10,] 0.018978966 0.015880793 0.009344760 0.005594973 0.002982074 0.007117232
#> [11,] 0.016119054 0.021007207 0.014642158 0.005493849 0.002620525 0.003658904
#> [12,] 0.017397100 0.013875385 0.011608377 0.008379110 0.007035868 0.007823690
#> [13,] 0.050869201 0.003275013 0.009015958 0.002961821 0.003411506 0.009016268
#> [14,] 0.024230117 0.030240872 0.018000306 0.012099019 0.013577436 0.011157667
#> [15,] 0.028795659 0.023996411 0.009920003 0.012005239 0.009409341 0.013458567
#> [16,] 0.017015117 0.005821691 0.014136540 0.009363490 0.010681441 0.011082490
#> [17,] 0.008128902 0.004540867 0.018517400 0.026961489 0.016145851 0.019431297
#> [18,] 0.013418792 0.012529619 0.007535866 0.002259465 0.012141000 0.007323157
#> [19,] 0.019073959 0.017132345 0.013229664 0.010489330 0.012274475 0.007835065
#> [20,] 0.017024751 0.017017948 0.005333105 0.008004345 0.006988026 0.007662130
#> [21,] 0.007878031 0.004107389 0.005515078 0.006406768 0.003532387 0.004718814
#> [22,] 0.014155277 0.011349842 0.009265443 0.003213333 0.005073724 0.007937708
#> [23,] 0.017847911 0.013032186 0.009753192 0.012045878 0.006802286 0.004542478
#> [24,] 0.012259828 0.016413717 0.003600718 0.008513189 0.011644518 0.008976238
#> [25,] 0.009811511 0.007777989 0.009114759 0.010043577 0.008991374 0.011306338
#>               V56          V57         V58         V59         V6           V60
#>  [1,] 0.007035750 0.0061310652 0.006325611 0.007633143 0.09581700  0.0089785326
#>  [2,] 0.011399516 0.0050642089 0.013932842 0.014257701 0.11686968  0.0094165849
#>  [3,] 0.008147333 0.0078320945 0.008720750 0.008574142 0.16331042  0.0059322553
#>  [4,] 0.010964209 0.0074327771 0.007410162 0.005791651 0.14104158  0.0040407804
#>  [5,] 0.014413987 0.0057492866 0.011249529 0.004710852 0.06183906  0.0069878699
#>  [6,] 0.015405072 0.0110153642 0.012547926 0.012623397 0.11011984  0.0058780192
#>  [7,] 0.003591632 0.0043273543 0.005768925 0.003010047 0.09093915  0.0023120931
#>  [8,] 0.013722117 0.0157272102 0.010073020 0.015315566 0.08772528  0.0115988543
#>  [9,] 0.007811162 0.0064143677 0.008990047 0.006716315 0.11980090  0.0049716385
#> [10,] 0.005811988 0.0013255379 0.006226241 0.005707370 0.08550663  0.0035727569
#> [11,] 0.004904708 0.0033375849 0.003079779 0.001508186 0.08758718  0.0039443420
#> [12,] 0.006609010 0.0094317479 0.009115800 0.007334267 0.09438586  0.0049540460
#> [13,] 0.001788628 0.0081664023 0.011250064 0.008264558 0.15236555  0.0031702684
#> [14,] 0.004880895 0.0051545769 0.005412812 0.005845328 0.13889863  0.0042712888
#> [15,] 0.016869930 0.0013992174 0.018131637 0.021075737 0.11689443  0.0162739303
#> [16,] 0.007690426 0.0061941181 0.004682192 0.006080673 0.06321188  0.0028898560
#> [17,] 0.005438018 0.0043865536 0.004798878 0.009628215 0.05970141  0.0095501501
#> [18,] 0.007467696 0.0068395671 0.005060458 0.006543592 0.09616737  0.0062336391
#> [19,] 0.006056130 0.0109267487 0.010115228 0.010997157 0.11375896  0.0092641467
#> [20,] 0.005949146 0.0057932992 0.011505118 0.006511986 0.06548513 -0.0002081323
#> [21,] 0.003574273 0.0046433481 0.005275793 0.006839919 0.06275109  0.0049737761
#> [22,] 0.006993217 0.0128992103 0.012332386 0.009812378 0.09769355  0.0096862613
#> [23,] 0.009082364 0.0065267850 0.002184688 0.003861759 0.08097899  0.0036891989
#> [24,] 0.005750593 0.0008319043 0.002053057 0.002346052 0.08238480  0.0033578849
#> [25,] 0.003843682 0.0050090097 0.006322637 0.004594033 0.10642398  0.0034129942
#>               V7         V8         V9
#>  [1,] 0.11240644 0.14137311 0.18154329
#>  [2,] 0.13128281 0.11326672 0.14445376
#>  [3,] 0.20834514 0.18520410 0.21744045
#>  [4,] 0.13870930 0.16690435 0.20367953
#>  [5,] 0.07568850 0.18734664 0.32657845
#>  [6,] 0.15143938 0.11571835 0.29642156
#>  [7,] 0.11236680 0.06960148 0.11430326
#>  [8,] 0.13942316 0.12383686 0.05942033
#>  [9,] 0.10593294 0.12691988 0.22141994
#> [10,] 0.09035062 0.17032230 0.22936983
#> [11,] 0.11495833 0.14187800 0.20239519
#> [12,] 0.14176706 0.18291486 0.23132335
#> [13,] 0.15466645 0.15327644 0.25657449
#> [14,] 0.15065021 0.21557543 0.32978366
#> [15,] 0.16804523 0.20712353 0.02866177
#> [16,] 0.08461552 0.07179917 0.08047865
#> [17,] 0.02544797 0.05422978 0.12252373
#> [18,] 0.12152725 0.07196432 0.08538316
#> [19,] 0.16866111 0.21373299 0.25078525
#> [20,] 0.06554160 0.05228737 0.05780556
#> [21,] 0.06310298 0.07292859 0.09942311
#> [22,] 0.11386078 0.10562877 0.06766765
#> [23,] 0.14052441 0.12637602 0.15013309
#> [24,] 0.09422460 0.04971502 0.04969896
#> [25,] 0.11109785 0.13006958 0.17355990
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696