Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.019934112 0.060138040 0.20032600 0.29134840 0.16551790 0.16583652
#>  [2,] 0.031811205 0.297136260 0.33475975 0.35446064 0.28422238 0.18338513
#>  [3,] 0.033277231 0.317009183 0.38864836 0.44192128 0.43973172 0.35255357
#>  [4,] 0.059435228 0.118704350 0.25599648 0.35574126 0.51368300 0.53017217
#>  [5,] 0.054179447 0.271143422 0.37376985 0.39459347 0.51201603 0.67836848
#>  [6,] 0.037599487 0.191150034 0.26711135 0.34429308 0.34564040 0.30346615
#>  [7,] 0.018764755 0.380526157 0.42872020 0.35931349 0.31782815 0.27229862
#>  [8,] 0.054860487 0.299703325 0.38542804 0.44938064 0.53687853 0.54184296
#>  [9,] 0.020081727 0.172431430 0.32120764 0.48177139 0.54906771 0.54977876
#> [10,] 0.041749046 0.236100593 0.31291393 0.31823257 0.35203867 0.36303025
#> [11,] 0.014511183 0.241473462 0.32368474 0.34153764 0.36919121 0.33312345
#> [12,] 0.031066039 0.350621470 0.40971352 0.43224849 0.36428478 0.37767079
#> [13,] 0.007988301 0.230379974 0.20623219 0.12223303 0.07189000 0.19326503
#> [14,] 0.038535754 0.527957372 0.36102858 0.12698946 0.26091845 0.38174375
#> [15,] 0.012814524 0.071901869 0.08162071 0.06959196 0.10724033 0.11661201
#> [16,] 0.025688648 0.073194461 0.04186567 0.09787178 0.10007143 0.09116932
#> [17,] 0.048484103 0.159432329 0.23376671 0.31949468 0.32871328 0.16818931
#> [18,] 0.011532025 0.188370961 0.15550612 0.11617702 0.23208972 0.31634666
#> [19,] 0.029943390 0.391106865 0.29654933 0.10089959 0.27809071 0.54393707
#> [20,] 0.016506426 0.128958192 0.12812115 0.13416980 0.11034168 0.14806951
#> [21,] 0.020615647 0.106809821 0.09126904 0.06991295 0.08266221 0.10868738
#> [22,] 0.022882247 0.305560659 0.38401494 0.42623517 0.42386290 0.50061120
#> [23,] 0.011061372 0.003410859 0.02256260 0.09037127 0.14126004 0.15974190
#> [24,] 0.025630258 0.099590205 0.24715446 0.36910816 0.30157405 0.27645769
#> [25,] 0.013700231 0.139933579 0.16439517 0.09665666 0.09543859 0.11260416
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.26246276 0.34228567 0.3407048 0.2213364 0.3214572 0.060849070 0.3201006
#>  [2,] 0.17839700 0.24048498 0.2624036 0.3356483 0.4006744 0.037378249 0.4609631
#>  [3,] 0.29418173 0.32476656 0.2961583 0.2607874 0.3918558 0.057307863 0.6163958
#>  [4,] 0.64604897 0.55271173 0.7138009 0.8307481 0.9607888 0.058386247 0.8829915
#>  [5,] 0.78677362 0.81608903 0.8189103 0.8826348 0.9579196 0.065208110 0.8712103
#>  [6,] 0.20117309 0.12514321 0.1821900 0.2382344 0.2943727 0.050178674 0.3081318
#>  [7,] 0.18395370 0.09290368 0.2471021 0.4612958 0.6811881 0.037263015 0.7955532
#>  [8,] 0.65152700 0.74831963 0.9125469 0.9220905 0.8710242 0.097326101 0.7679361
#>  [9,] 0.62030879 0.74473796 0.8793135 0.9778608 0.9766296 0.032528043 0.9267313
#> [10,] 0.29467217 0.26220589 0.2908491 0.2471786 0.3702952 0.059671611 0.4797986
#> [11,] 0.17978865 0.07397620 0.2339955 0.3110950 0.3494168 0.028204062 0.4430207
#> [12,] 0.41625197 0.47489270 0.5667907 0.5881657 0.6264193 0.041381136 0.6858606
#> [13,] 0.35865974 0.65041274 0.7921132 0.8155184 0.8313707 0.008751942 0.7652830
#> [14,] 0.39739912 0.42628510 0.1806242 0.1845452 0.2874328 0.067110796 0.8970180
#> [15,] 0.14800122 0.21550123 0.2729677 0.3339941 0.4219722 0.014453486 0.4183838
#> [16,] 0.12925882 0.11618200 0.1360276 0.1581792 0.2545234 0.050969701 0.2724367
#> [17,] 0.25050810 0.35455610 0.4683127 0.5492108 0.5308181 0.055730932 0.3700022
#> [18,] 0.36947516 0.24833310 0.1048453 0.1652682 0.2223495 0.021359202 0.2530851
#> [19,] 0.67818454 0.87491573 0.9175299 0.8199863 0.5455009 0.071355973 0.4132840
#> [20,] 0.10444009 0.16345874 0.2842474 0.3309695 0.3130275 0.014331106 0.4894491
#> [21,] 0.05058579 0.14183013 0.1609991 0.3431995 0.3381379 0.020408857 0.3230327
#> [22,] 0.66972222 0.80397225 0.9469340 0.9388792 0.8991058 0.031743432 0.8247315
#> [23,] 0.11123478 0.10869001 0.2056019 0.1789592 0.1124928 0.023846415 0.1092520
#> [24,] 0.30509956 0.45431335 0.5996042 0.5301737 0.4983463 0.056273014 0.6349840
#> [25,] 0.21670586 0.18779771 0.1611969 0.2082267 0.2487204 0.019442592 0.3178443
#>             V21       V22       V23       V24        V25       V26        V27
#>  [1,] 0.5158597 0.5052609 0.5587179 0.7118704 0.58055267 0.5959883 0.77961999
#>  [2,] 0.6354234 0.7879002 0.8595462 0.9009746 0.94617998 1.0108207 1.01215951
#>  [3,] 0.7114808 0.6881162 0.6368895 0.6574431 0.65299194 0.7709160 0.90602626
#>  [4,] 0.9787616 1.0328135 1.0170306 0.8812500 0.62782409 0.4502171 0.31452683
#>  [5,] 0.8668640 0.9068599 0.8962441 0.7814307 0.64274603 0.6338709 0.49673983
#>  [6,] 0.3936257 0.5019112 0.6411061 0.7495898 0.74291769 0.7136395 0.86218304
#>  [7,] 0.7470179 0.5323699 0.3526990 0.4259496 0.56825324 0.8046667 0.98508615
#>  [8,] 0.7338242 0.5264503 0.2661847 0.1997962 0.21843898 0.2268330 0.24325619
#>  [9,] 0.7833576 0.6864147 0.6466244 0.5668989 0.35608173 0.1087513 0.06169943
#> [10,] 0.6651195 0.7393089 0.7057892 0.7154272 0.68955565 0.6401964 0.67379486
#> [11,] 0.4575809 0.4394394 0.4538458 0.5492703 0.66155052 0.7319736 0.93036637
#> [12,] 0.7978186 0.9671535 1.0040251 0.9324324 0.80683765 0.7709939 0.76067878
#> [13,] 0.7887702 0.9598122 0.8926055 0.4405813 0.17453613 0.2967861 0.56009524
#> [14,] 0.9850451 0.3486020 0.1595352 0.1660087 0.14257530 0.3595036 0.28432034
#> [15,] 0.4074994 0.3861333 0.3582872 0.4464270 0.46827087 0.5155469 0.68409014
#> [16,] 0.2070073 0.1726106 0.2845040 0.3253311 0.31322184 0.2411465 0.23665902
#> [17,] 0.3164890 0.2278356 0.2165048 0.2136614 0.27162189 0.3030831 0.24913046
#> [18,] 0.2529402 0.3431075 0.5829250 0.7541565 0.71698095 0.7237162 0.73220790
#> [19,] 0.5165207 0.5968875 0.7585783 0.8868961 0.84534053 0.7323924 0.75338190
#> [20,] 0.6180549 0.7487186 0.8675108 0.8713985 0.91521646 0.9607467 0.93043408
#> [21,] 0.2420254 0.1941855 0.1884434 0.2406639 0.09192075 0.2119253 0.41733616
#> [22,] 0.8001801 0.8293875 0.6154463 0.5200112 0.67309989 0.7180875 0.59563039
#> [23,] 0.2088656 0.4462503 0.4643367 0.5573811 0.69365875 0.7866709 0.76293986
#> [24,] 0.4383369 0.2819507 0.4227036 0.8855391 0.93220297 0.5801764 0.43349127
#> [25,] 0.3569966 0.6012096 0.7316507 0.6912840 0.66599147 0.7178093 0.65618400
#>              V28       V29          V3       V30       V31        V32       V33
#>  [1,] 0.75594350 0.6415491  0.07313352 0.8573869 0.6956410 0.56198320 0.3695583
#>  [2,] 0.99932178 0.8775671  0.02596644 0.6698204 0.5629084 0.45157554 0.3360179
#>  [3,] 0.92308053 0.7983132  0.06059701 0.6591179 0.5891207 0.58320128 0.4330749
#>  [4,] 0.25463593 0.3390989  0.09600187 0.4555381 0.3186422 0.05474565 0.2532016
#>  [5,] 0.39908879 0.4778263  0.07082162 0.3697385 0.2944708 0.14936601 0.1241918
#>  [6,] 1.05374644 1.0043705  0.05742382 0.6938520 0.7018255 0.59564001 0.4119132
#>  [7,] 0.97361697 0.8736611  0.04395213 0.7715442 0.5928206 0.53146874 0.7200280
#>  [8,] 0.23184963 0.1388532  0.10959110 0.1799107 0.2206759 0.26169734 0.3224212
#>  [9,] 0.07175185 0.1031433  0.03666728 0.1960824 0.3111408 0.41617197 0.3667547
#> [10,] 0.75738554 0.9069472  0.05881590 0.9660678 0.8878344 0.70445294 0.6118147
#> [11,] 1.12008876 1.0713833  0.01519788 0.7400583 0.6260458 0.49389803 0.4061862
#> [12,] 0.61759620 0.5293816  0.05730369 0.4534570 0.3091150 0.26996565 0.2639542
#> [13,] 0.74505187 0.8990872 -0.01093520 0.8316062 0.4475007 0.35877675 0.4023703
#> [14,] 0.52689615 0.4713049  0.09141478 0.6998093 0.2373794 0.30435204 0.5829944
#> [15,] 0.89253208 1.0000026  0.01452417 0.9257696 0.8517434 0.78574048 0.7401121
#> [16,] 0.44367954 0.6201982  0.03788617 0.7812000 0.7917341 0.67161801 0.7384336
#> [17,] 0.45373389 0.7144080  0.06682538 0.7647868 0.7491986 0.70899934 0.6326855
#> [18,] 0.86252432 0.9610214  0.04054879 0.8812789 0.7728597 0.64410230 0.6389938
#> [19,] 0.60902850 0.3125850  0.07666238 0.3502300 0.5613278 0.48812624 0.2607047
#> [20,] 0.83597906 0.7639694  0.02060022 0.7035016 0.6924209 0.63040485 0.5920418
#> [21,] 0.66445690 0.8213789  0.01963779 0.7577523 0.5345250 0.22755190 0.4832087
#> [22,] 0.51772814 0.6164688  0.04312377 0.7096162 0.6254819 0.41206281 0.3650033
#> [23,] 0.78185717 0.8679773  0.02468503 0.8423827 0.7425265 0.75362950 0.7977351
#> [24,] 0.71772795 0.8126877  0.05479937 0.4872883 0.5490513 0.70287994 0.6309590
#> [25,] 0.68706090 0.7330357  0.02541030 0.6959158 0.5214482 0.43782782 0.6798577
#>             V34       V35        V36        V37       V38       V39         V4
#>  [1,] 0.4310841 0.2749624 0.06164440 0.26304664 0.3911872 0.1921114 0.06551722
#>  [2,] 0.1906175 0.1129095 0.01965489 0.06798934 0.1465403 0.1564883 0.03410733
#>  [3,] 0.3231868 0.1915379 0.29449028 0.30141328 0.3867840 0.5279737 0.04350380
#>  [4,] 0.5174118 0.5971470 0.50835577 0.40600283 0.4266445 0.3253250 0.12341272
#>  [5,] 0.1832735 0.2097309 0.17710657 0.17328385 0.1691716 0.3146904 0.07936533
#>  [6,] 0.2976004 0.1353083 0.06515856 0.27009765 0.2086491 0.2854441 0.07151433
#>  [7,] 0.6927858 0.5548771 0.39107930 0.38736881 0.3801938 0.3155544 0.07252587
#>  [8,] 0.3452391 0.4365150 0.15512084 0.13530453 0.2667081 0.2743990 0.14450054
#>  [9,] 0.4768374 0.4518982 0.34202479 0.29320451 0.4121358 0.3179139 0.04221402
#> [10,] 0.5820230 0.4548349 0.31304890 0.21892862 0.1640804 0.2461237 0.06385537
#> [11,] 0.3449674 0.1523391 0.07701458 0.17411809 0.1233162 0.1452257 0.05811810
#> [12,] 0.2043880 0.1942767 0.23379577 0.32963617 0.3131511 0.3124493 0.06331523
#> [13,] 0.4056164 0.2695110 0.12412348 0.15295377 0.3740144 0.3813158 0.01045150
#> [14,] 0.8021105 0.8102863 0.86269407 0.79341413 0.8918985 0.6858201 0.11339917
#> [15,] 0.6255298 0.5565561 0.53754156 0.50293404 0.3077176 0.1201810 0.01330081
#> [16,] 0.9177594 0.9932753 0.92208422 0.62266153 0.4128562 0.3579311 0.01659528
#> [17,] 0.6557819 0.7022787 0.79223892 0.85191855 0.9611845 0.9243813 0.02700889
#> [18,] 0.5591277 0.4313008 0.38207338 0.23065339 0.1441394 0.1944009 0.03729580
#> [19,] 0.3229905 0.7045344 0.71588731 0.57484934 0.5184350 0.3548888 0.06294385
#> [20,] 0.5783779 0.5772688 0.51746212 0.45994565 0.3473929 0.1643690 0.02908826
#> [21,] 0.4890651 0.2605738 0.35407654 0.49097182 0.5785225 0.6291305 0.02937684
#> [22,] 0.3823876 0.2338578 0.18079038 0.28253305 0.2295336 0.2174006 0.04677318
#> [23,] 0.6582872 0.6314568 0.63297716 0.55472381 0.5169618 0.4666122 0.02747102
#> [24,] 0.8942934 0.8326893 0.76970753 0.79120012 0.4937237 0.3621571 0.07671597
#> [25,] 0.9162541 0.8491506 0.65449437 0.36303453 0.2977384 0.2306935 0.03069575
#>              V40        V41        V42        V43        V44        V45
#>  [1,] 0.39116008 0.49204837 0.22765546 0.29764010 0.33753412 0.25302020
#>  [2,] 0.12691431 0.11162663 0.09285116 0.17887955 0.18708329 0.15933481
#>  [3,] 0.43706861 0.37626984 0.50800089 0.48488679 0.44361562 0.51977863
#>  [4,] 0.13698748 0.14307775 0.31087929 0.37682873 0.23261680 0.13001518
#>  [5,] 0.30443186 0.16792789 0.16739567 0.29699865 0.20434248 0.10210562
#>  [6,] 0.29973924 0.25581871 0.33271099 0.15831131 0.05576357 0.10316144
#>  [7,] 0.26547898 0.34970467 0.34925360 0.32685437 0.21248630 0.11444734
#>  [8,] 0.13761998 0.08414428 0.12460082 0.11882064 0.06540518 0.08007186
#>  [9,] 0.20095381 0.22058677 0.22123042 0.16352563 0.06562166 0.06234179
#> [10,] 0.33897357 0.51572232 0.54902153 0.37205917 0.18535945 0.31448736
#> [11,] 0.31192778 0.24670607 0.35016709 0.22420901 0.14840281 0.24867680
#> [12,] 0.33762802 0.31147378 0.34796975 0.29527837 0.22007074 0.19508034
#> [13,] 0.27887530 0.11066508 0.16009703 0.35634123 0.32652075 0.13798190
#> [14,] 0.48413362 0.45436587 0.34326304 0.20657755 0.19389993 0.32914153
#> [15,] 0.07715424 0.07106499 0.17596553 0.26067735 0.30129681 0.25776069
#> [16,] 0.38027993 0.40827055 0.42327812 0.22576084 0.21097565 0.14559722
#> [17,] 0.80086934 0.75106294 0.66914347 0.54482819 0.42121750 0.28529175
#> [18,] 0.13557181 0.13938860 0.20165530 0.26469642 0.17230588 0.03485605
#> [19,] 0.30094199 0.19108057 0.22052204 0.27270027 0.25704791 0.22112025
#> [20,] 0.17887940 0.19124569 0.14786599 0.12298383 0.05804313 0.03164919
#> [21,] 0.48735114 0.24354627 0.18986573 0.13790930 0.06926317 0.04134194
#> [22,] 0.31410077 0.24916660 0.22546990 0.14627163 0.12460671 0.18485787
#> [23,] 0.33673508 0.18570677 0.07615105 0.11668617 0.12520748 0.15035063
#> [24,] 0.35174038 0.28403224 0.27247590 0.07535893 0.17335501 0.15464323
#> [25,] 0.35304491 0.34713793 0.38745223 0.39291416 0.24850734 0.11595443
#>              V46        V47          V48        V49         V5         V50
#>  [1,] 0.14646487 0.07503842 1.188150e-01 0.09311302 0.02112545 0.017816760
#>  [2,] 0.15387782 0.15503273 1.177529e-01 0.07167313 0.06242646 0.019800484
#>  [3,] 0.32872624 0.21114629 1.863538e-01 0.11569319 0.07902598 0.033930803
#>  [4,] 0.06127071 0.07472357 2.717292e-05 0.03250123 0.14774753 0.010702105
#>  [5,] 0.12478475 0.09973906 6.478721e-02 0.05121791 0.07617648 0.021437569
#>  [6,] 0.22494546 0.27853341 1.955865e-01 0.06297019 0.10103678 0.012761945
#>  [7,] 0.09002610 0.07429770 6.244602e-02 0.02258410 0.15455760 0.012773615
#>  [8,] 0.07078956 0.08885523 9.413991e-02 0.05856330 0.14582964 0.029033097
#>  [9,] 0.04786048 0.03732238 2.353144e-02 0.02025269 0.06358980 0.010725070
#> [10,] 0.26495125 0.16552657 1.234298e-01 0.07197665 0.08967610 0.019171335
#> [11,] 0.13596552 0.20603399 1.536376e-01 0.07162794 0.10680264 0.007377401
#> [12,] 0.15544816 0.12991642 8.452041e-02 0.04538101 0.10239215 0.019170271
#> [13,] 0.15120870 0.13584461 7.894601e-02 0.05617397 0.04523347 0.020015902
#> [14,] 0.23693267 0.14872323 2.137787e-01 0.10640533 0.05794697 0.030460624
#> [15,] 0.18099658 0.10241752 9.384921e-02 0.05605160 0.02271632 0.016886903
#> [16,] 0.10964549 0.13614219 1.064065e-01 0.05563100 0.03749458 0.015132930
#> [17,] 0.26481339 0.22625806 1.653603e-01 0.09396979 0.05580125 0.035146931
#> [18,] 0.03527168 0.03776015 3.968378e-02 0.01573625 0.02359101 0.016087808
#> [19,] 0.12316407 0.10303272 9.094945e-02 0.05335889 0.09611627 0.027438533
#> [20,] 0.04251974 0.03247672 3.850417e-02 0.01530983 0.06223046 0.008655004
#> [21,] 0.02884680 0.07635918 7.373674e-02 0.02445036 0.04438695 0.010952529
#> [22,] 0.19072052 0.18975724 1.261742e-01 0.05587941 0.09659915 0.020398779
#> [23,] 0.09664032 0.06220553 3.876080e-02 0.01529596 0.01964622 0.010201267
#> [24,] 0.15762670 0.11584673 6.727338e-02 0.03849576 0.12673782 0.009674906
#> [25,] 0.07336123 0.08769723 6.722359e-02 0.02120675 0.05807833 0.024858172
#>               V51         V52         V53         V54          V55         V56
#>  [1,] 0.024982368 0.024741319 0.015541820 0.020035628  0.019848363 0.017856268
#>  [2,] 0.019754878 0.010942461 0.006294096 0.008070222  0.004566676 0.006808001
#>  [3,] 0.032642446 0.019813192 0.008378988 0.010624769  0.012556723 0.006310042
#>  [4,] 0.016859789 0.009274228 0.001687858 0.030027897  0.013550644 0.012258790
#>  [5,] 0.030072073 0.011078425 0.009215209 0.023321978  0.010203551 0.009405706
#>  [6,] 0.023881743 0.019039221 0.009863910 0.009270520  0.007740817 0.005079564
#>  [7,] 0.012130663 0.012786788 0.010955339 0.010678558  0.002166606 0.004442940
#>  [8,] 0.026203968 0.030298290 0.018497489 0.011097539  0.017165661 0.012089171
#>  [9,] 0.011718981 0.013792632 0.011064446 0.020656800  0.021914872 0.009488101
#> [10,] 0.019204723 0.020069222 0.009740800 0.011137141  0.005636306 0.006952599
#> [11,] 0.029291364 0.009472575 0.011512850 0.006435591 -0.001208565 0.008465665
#> [12,] 0.012557812 0.014267447 0.015170485 0.013749802  0.005409605 0.011926557
#> [13,] 0.012133950 0.016675745 0.006692213 0.001226720  0.002335203 0.004477178
#> [14,] 0.037909859 0.020703733 0.018394074 0.033302804  0.016748883 0.028991199
#> [15,] 0.010590829 0.006928453 0.002136836 0.009269931  0.005181115 0.002519089
#> [16,] 0.013875982 0.003616544 0.009834209 0.011252750  0.007720518 0.007100221
#> [17,] 0.019561654 0.014322059 0.017904129 0.005982076  0.008744862 0.007490132
#> [18,] 0.017320332 0.015138983 0.013732262 0.005267565  0.005963646 0.006509974
#> [19,] 0.008116129 0.018348536 0.006818518 0.017887495  0.011728542 0.003472281
#> [20,] 0.004968056 0.004730958 0.007954278 0.006283213  0.006337802 0.005233342
#> [21,] 0.005861917 0.018480791 0.025088349 0.015849302  0.019114878 0.006295474
#> [22,] 0.009764955 0.010144495 0.008092563 0.008303262  0.010595386 0.009905014
#> [23,] 0.003835482 0.006011281 0.012485789 0.006589193  0.010033826 0.009432211
#> [24,] 0.017402129 0.006730546 0.013992341 0.007202818  0.018314199 0.008589174
#> [25,] 0.027238384 0.013947796 0.013245843 0.012664057  0.005195623 0.011499371
#>                V57         V58         V59         V6         V60         V7
#>  [1,]  0.013652413 0.011157987 0.008940370 0.10369053 0.011323854 0.14379054
#>  [2,]  0.006779858 0.008803795 0.008578111 0.14292872 0.007145477 0.16104642
#>  [3,]  0.006396610 0.006756569 0.006355786 0.12676031 0.005507625 0.12036636
#>  [4,]  0.029051092 0.041486508 0.021821558 0.16965913 0.008810597 0.20097502
#>  [5,]  0.010426298 0.013262829 0.011613951 0.14547303 0.009552531 0.21080953
#>  [6,]  0.007986834 0.007864959 0.004477679 0.15144862 0.005465828 0.13676480
#>  [7,]  0.002485975 0.005184228 0.008277404 0.09562684 0.003099733 0.09639736
#>  [8,]  0.005745835 0.015252030 0.020558160 0.08184576 0.014130794 0.10288587
#>  [9,]  0.003955099 0.010562181 0.006834560 0.05001091 0.002863739 0.09513272
#> [10,]  0.008951965 0.008785412 0.006842160 0.16450425 0.009419399 0.10771107
#> [11,]  0.001673819 0.007165291 0.002595180 0.15885581 0.001205442 0.16453640
#> [12,]  0.008155730 0.009597045 0.008757061 0.11107092 0.007864158 0.11975626
#> [13,] -0.002894958 0.002599368 0.003085424 0.09099820 0.003747818 0.08631002
#> [14,]  0.017595977 0.016214867 0.020663536 0.10834528 0.012146960 0.19100253
#> [15,]  0.003454972 0.003744195 0.007145582 0.08604557 0.003504534 0.09796333
#> [16,]  0.010059934 0.015441276 0.013719572 0.06305138 0.002618518 0.07278744
#> [17,]  0.001345506 0.017100485 0.011990429 0.08541007 0.008555963 0.11682709
#> [18,]  0.006161284 0.005241015 0.014189903 0.06748632 0.008593241 0.05927470
#> [19,]  0.008370026 0.005714217 0.009963306 0.19865228 0.006825509 0.22394304
#> [20,]  0.005555433 0.005393244 0.004006794 0.10146682 0.006030842 0.09974179
#> [21,]  0.005138950 0.005554694 0.010043476 0.06051858 0.009094607 0.03116018
#> [22,]  0.010616468 0.005172868 0.004697828 0.13160866 0.005109941 0.15072622
#> [23,]  0.005166354 0.002491606 0.007111104 0.02700292 0.005252220 0.05823833
#> [24,]  0.013704327 0.006222820 0.015717070 0.20522951 0.006343281 0.19589014
#> [25,]  0.013971832 0.004789385 0.005263763 0.07590648 0.005590828 0.11075019
#>               V8          V9
#>  [1,] 0.12213672  0.16168159
#>  [2,] 0.17158439  0.22801029
#>  [3,] 0.18124467  0.25125379
#>  [4,] 0.07297421  0.08445871
#>  [5,] 0.26911074  0.27612796
#>  [6,] 0.12728439  0.11827209
#>  [7,] 0.17809412  0.40599540
#>  [8,] 0.05466930  0.15198205
#>  [9,] 0.11941814  0.08287620
#> [10,] 0.10564167  0.19838757
#> [11,] 0.13633711  0.19973620
#> [12,] 0.17473771  0.27932034
#> [13,] 0.03561912  0.13461705
#> [14,] 0.42272662  0.62710038
#> [15,] 0.05034104  0.08006651
#> [16,] 0.04214767  0.06243478
#> [17,] 0.16022965  0.16465826
#> [18,] 0.10733889  0.16103131
#> [19,] 0.22887888  0.32350368
#> [20,] 0.09991296  0.12588663
#> [21,] 0.07256774  0.13742157
#> [22,] 0.18243322  0.27140596
#> [23,] 0.03273569 -0.01445331
#> [24,] 0.17371296  0.12627501
#> [25,] 0.12303444  0.12347938
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3768116