Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12          V13       V14
#>  [1,] 0.010395875 0.34841114 0.33674463 0.34840008  0.247336285 0.1670695
#>  [2,] 0.016233970 0.25990281 0.30175315 0.29434689  0.275762113 0.1864364
#>  [3,] 0.088772305 0.65875703 0.50262014 0.24378495  0.135773362 0.4022557
#>  [4,] 0.048495444 0.19875965 0.19140973 0.23544094  0.293149022 0.2908951
#>  [5,] 0.042974105 0.21101708 0.31419893 0.38153920  0.462786385 0.5579082
#>  [6,] 0.033957828 0.17618100 0.24153552 0.29910699  0.299482024 0.2596809
#>  [7,] 0.063980576 0.33510305 0.21289433 0.14142533  0.400633029 0.5607933
#>  [8,] 0.019192898 0.23864120 0.27750274 0.30028941  0.319270273 0.3350534
#>  [9,] 0.078210511 0.50572458 0.34648742 0.14885955  0.227819320 0.3306769
#> [10,] 0.009670737 0.20460734 0.32294600 0.30973771  0.272568287 0.2625660
#> [11,] 0.016558816 0.34815052 0.47748959 0.44021182  0.336906846 0.2698393
#> [12,] 0.026133168 0.43221471 0.49337810 0.49554299  0.489403994 0.4017430
#> [13,] 0.024723631 0.40997067 0.47989160 0.49472547  0.507028248 0.3978158
#> [14,] 0.009300000 0.13600000 0.21320000 0.37380000  0.373800000 0.2673000
#> [15,] 0.019206189 0.08750678 0.05429955 0.02488023  0.123159709 0.2912062
#> [16,] 0.006530243 0.04592019 0.06986620 0.08469529  0.135135518 0.1428530
#> [17,] 0.022712124 0.08117082 0.10175585 0.15251658  0.127578355 0.0915070
#> [18,] 0.005086534 0.06082305 0.01098724 0.02925471  0.228047924 0.2709599
#> [19,] 0.016424588 0.39710983 0.30824556 0.15711764  0.294364956 0.5208543
#> [20,] 0.013895393 0.11686111 0.12706545 0.14339442  0.193512094 0.1883743
#> [21,] 0.012218085 0.02949167 0.13995035 0.21824576  0.290260486 0.3700446
#> [22,] 0.012819069 0.08093518 0.08689390 0.07374186  0.111749091 0.1193058
#> [23,] 0.005762821 0.11812070 0.08738477 0.19372743 -0.003544471 0.2159785
#> [24,] 0.015140185 0.13283917 0.12427877 0.10234166  0.118719559 0.1699077
#> [25,] 0.052744812 0.28002170 0.25632230 0.08872266  0.149343596 0.4624379
#>              V15       V16        V17       V18       V19          V2       V20
#>  [1,] 0.13071791 0.2330114 0.26470433 0.3215353 0.2343591 0.027276263 0.1377311
#>  [2,] 0.14519392 0.1359740 0.17594295 0.2439320 0.2519026 0.035071905 0.2591962
#>  [3,] 0.46595491 0.4630561 0.26438483 0.2893658 0.4652515 0.088877260 0.9216823
#>  [4,] 0.32903828 0.3789317 0.36769853 0.1828181 0.3342911 0.079932509 0.3874169
#>  [5,] 0.63876376 0.7168545 0.81475278 0.8604733 0.8846751 0.056036714 0.8422333
#>  [6,] 0.17711049 0.1446398 0.19622576 0.2384731 0.2899209 0.039516510 0.3023681
#>  [7,] 0.23429534 0.4298028 0.59128740 0.5789461 0.8185809 0.062798873 0.8804251
#>  [8,] 0.35258707 0.3994247 0.48174309 0.5612061 0.6584717 0.044142787 0.7294513
#>  [9,] 0.41654006 0.5029049 0.21396925 0.1703134 0.2001325 0.077942544 0.8518904
#> [10,] 0.19638976 0.1726636 0.22997880 0.3291360 0.4443692 0.033918523 0.4666395
#> [11,] 0.28388594 0.4495233 0.44768862 0.4519933 0.5133148 0.032390518 0.5625823
#> [12,] 0.27830551 0.2225790 0.28862115 0.3585463 0.5267798 0.010529819 0.6826561
#> [13,] 0.28136927 0.3007751 0.30078243 0.2953539 0.4449695 0.034816471 0.6847433
#> [14,] 0.23330000 0.5367000 0.73120000 0.7659000 0.6271000 0.026900000 0.4395000
#> [15,] 0.53067026 0.7670526 0.82519533 0.8017123 0.8481872 0.012627456 0.8726682
#> [16,] 0.10075065 0.1273355 0.14209238 0.1491938 0.1221434 0.005016191 0.1076758
#> [17,] 0.07885806 0.1641642 0.20261784 0.3210349 0.2978140 0.026429348 0.2912923
#> [18,] 0.19740062 0.2412525 0.12081068 0.3160719 0.7170785 0.021452612 0.9355492
#> [19,] 0.65465489 0.9234891 0.94431232 0.9074452 0.6752797 0.048117441 0.5651631
#> [20,] 0.22158117 0.2817645 0.34231954 0.3091209 0.2458233 0.014442096 0.3067741
#> [21,] 0.48113513 0.5332564 0.61083221 0.7111489 0.7424135 0.005402281 0.7765807
#> [22,] 0.15449043 0.2304687 0.29012434 0.3492014 0.4183147 0.014868093 0.4124830
#> [23,] 0.15544195 0.2001826 0.04668365 0.2205793 0.4180505 0.015929628 0.4216177
#> [24,] 0.17383961 0.2149893 0.26928224 0.2794996 0.2690458 0.016153116 0.3717859
#> [25,] 0.64789245 0.6966171 0.74873548 0.6199732 0.4109083 0.068424585 0.3191786
#>              V21       V22       V23        V24       V25       V26       V27
#>  [1,] 0.08405105 0.1578125 0.3866059 0.53197901 0.6531657 0.8190631 0.9438563
#>  [2,] 0.34483647 0.4976763 0.6478144 0.77465884 0.8069723 0.8398251 0.9112519
#>  [3,] 1.00815338 0.4130863 0.1060926 0.05920361 0.2438233 0.4542721 0.2650675
#>  [4,] 0.47144609 0.4673388 0.4630917 0.51826675 0.5103032 0.5057307 0.6199880
#>  [5,] 0.79745316 0.7292564 0.6350417 0.56323272 0.4814809 0.4132350 0.3570551
#>  [6,] 0.36153962 0.4307823 0.5705963 0.72175525 0.7782714 0.7981881 0.8779518
#>  [7,] 0.53077605 0.6432565 0.8413394 0.67977875 0.7897651 0.8831958 0.8253848
#>  [8,] 0.80831861 0.8540262 0.8863575 0.95231952 0.9512607 0.9517758 0.9410003
#>  [9,] 0.99779059 0.3829884 0.1055470 0.18289058 0.2051548 0.3732166 0.2898316
#> [10,] 0.42436169 0.3259911 0.3060590 0.43537311 0.5620163 0.7639957 0.8958776
#> [11,] 0.75903383 0.9922225 1.0154395 0.98601434 0.9513409 0.8201075 0.6201318
#> [12,] 0.74736040 0.8503485 0.8539823 0.87185180 0.9010244 0.9453863 0.9702546
#> [13,] 0.79515565 0.7438631 0.6438948 0.67484823 0.6723336 0.8329121 0.9519127
#> [14,] 0.43300000 0.4326000 0.5544000 0.73600000 0.8589000 0.8989000 0.9420000
#> [15,] 0.94912311 0.9259850 0.9064719 0.95016463 0.9397531 0.9517092 0.9557422
#> [16,] 0.16926298 0.3540277 0.4464352 0.57693117 0.6692201 0.7677339 0.7948337
#> [17,] 0.23822076 0.1177464 0.1516019 0.15797024 0.0232396 0.1726032 0.3939947
#> [18,] 0.85642123 0.8584708 0.5025731 0.39042290 0.6352117 0.8097089 0.9604544
#> [19,] 0.60515008 0.5804273 0.6953603 0.79057821 0.8131905 0.7703659 0.7685976
#> [20,] 0.42092276 0.4453505 0.5361457 0.67596617 0.7947289 0.7968004 0.7933570
#> [21,] 0.89738258 0.9574527 0.9038330 0.89059616 0.8607791 0.8320882 0.7511663
#> [22,] 0.40941704 0.3833445 0.3452252 0.43147264 0.4660068 0.5255923 0.7093623
#> [23,] 0.26843237 0.3674395 0.5776312 0.47848806 0.2913728 0.5427088 0.5246897
#> [24,] 0.48900331 0.6371202 0.7526024 0.77634798 0.7828766 0.8266100 0.8220167
#> [25,] 0.55154930 0.7799171 0.8580923 0.90433184 0.8438174 0.6616903 0.6336243
#>             V28       V29            V3       V30       V31       V32       V33
#>  [1,] 0.9669607 0.9657793  0.0208981140 1.0236430 0.9493170 0.8309597 0.7162319
#>  [2,] 0.9872918 0.8591859  0.0280821243 0.6486077 0.6710823 0.5470488 0.3667411
#>  [3,] 0.5550263 0.6455323  0.0503571434 0.8406599 0.4546862 0.4050983 0.5502478
#>  [4,] 0.7935462 0.8421126  0.0743992556 0.8652167 0.8342061 0.6936362 0.4422085
#>  [5,] 0.2912573 0.2819882  0.0630944691 0.2853172 0.2670885 0.2729633 0.3040872
#>  [6,] 0.9744851 0.9050257  0.0427282869 0.6288093 0.6067248 0.4664617 0.3391443
#>  [7,] 0.6833479 0.3371482  0.0712450882 0.6373753 0.7139311 0.8231649 0.6749712
#>  [8,] 0.8775224 0.7640436  0.0404887908 0.5920795 0.4100685 0.3184530 0.2334662
#>  [9,] 0.3990475 0.3940143  0.0738391501 0.5666258 0.2156861 0.3150205 0.5051541
#> [10,] 0.9243966 0.8593234  0.0366168927 0.6663009 0.4994292 0.5018572 0.5769433
#> [11,] 0.5180002 0.3544606  0.0224009254 0.3120078 0.2998388 0.3775663 0.4689854
#> [12,] 0.7879312 0.5005135  0.0177178326 0.3869496 0.3721215 0.4712150 0.4917599
#> [13,] 0.9496855 0.8364403  0.0527872062 0.5910657 0.4830691 0.5643835 0.4905705
#> [14,] 0.9401000 0.9379000  0.0217000000 0.8575000 0.7284000 0.6700000 0.7547000
#> [15,] 0.7661769 0.4478112  0.0230090127 0.3321505 0.3885842 0.4004341 0.3419892
#> [16,] 0.8765733 0.9302605  0.0165777465 0.8600075 0.7123120 0.6573984 0.6777547
#> [17,] 0.6354796 0.7708933  0.0249169195 0.7373392 0.6440806 0.4156859 0.5732440
#> [18,] 0.8259239 0.5607108  0.0013714758 0.3839475 0.3862776 0.3261756 0.3591884
#> [19,] 0.5781034 0.2489230  0.0532950423 0.2334214 0.5487641 0.6522858 0.3460297
#> [20,] 0.8054326 0.6351195  0.0158623932 0.4420753 0.4002017 0.3073517 0.2183824
#> [21,] 0.5701475 0.6384030 -0.0007455183 0.7564029 0.5651057 0.2528274 0.3082829
#> [22,] 0.9073313 0.9966529  0.0143419765 0.9235033 0.8491248 0.7779413 0.7327953
#> [23,] 0.2216245 0.1893096  0.0626359400 0.3570162 0.6495050 0.7632701 0.6178519
#> [24,] 0.8288201 0.8545073  0.0182836801 0.8315856 0.7659188 0.6675973 0.6415747
#> [25,] 0.5890593 0.4279461  0.0845996639 0.4884146 0.4445861 0.2361087 0.2233180
#>             V34        V35        V36         V37         V38       V39
#>  [1,] 0.6626648 0.67919967 0.72331171 0.716494707 0.579613912 0.4851997
#>  [2,] 0.1808946 0.13331546 0.04950554 0.154739919 0.125521509 0.1901709
#>  [3,] 0.7548962 0.81529855 0.75021614 0.738157582 0.835842888 0.7577719
#>  [4,] 0.4110253 0.43592142 0.28002144 0.238047975 0.338778724 0.3220782
#>  [5,] 0.3260423 0.35263685 0.27867008 0.297442060 0.324826653 0.3034774
#>  [6,] 0.2688795 0.18050121 0.10882000 0.225365284 0.146516517 0.2209975
#>  [7,] 0.5346932 0.30541744 0.07722478 0.089609788 0.379281570 0.5925863
#>  [8,] 0.1833504 0.18369233 0.17226365 0.196142361 0.210722159 0.2079738
#>  [9,] 0.6656737 0.71404170 0.78438918 0.758687885 0.817454942 0.6781180
#> [10,] 0.4603045 0.28521899 0.19342570 0.217704274 0.201823425 0.2138014
#> [11,] 0.4763599 0.32596907 0.24035204 0.204340688 0.214097686 0.4265298
#> [12,] 0.4150459 0.20241129 0.35071900 0.437410160 0.338550747 0.3190211
#> [13,] 0.3011252 0.08451061 0.25465839 0.270827072 0.372908154 0.5736739
#> [14,] 0.8773000 0.99190000 0.99220000 0.941900000 0.838800000 0.6605000
#> [15,] 0.2709908 0.21887144 0.11138006 0.222691727 0.185158036 0.1007188
#> [16,] 0.5804501 0.50604537 0.43205192 0.311158460 0.306920929 0.3309773
#> [17,] 0.6827154 0.59952006 0.64694581 0.664734400 0.680512183 0.6816218
#> [18,] 0.3158646 0.26546705 0.22213127 0.334691839 0.430187651 0.3639480
#> [19,] 0.4933065 0.82703150 0.71467975 0.494116484 0.492401939 0.4939414
#> [20,] 0.2208737 0.28567496 0.30384483 0.236963366 0.217708985 0.1811369
#> [21,] 0.3573128 0.15200183 0.05485409 0.007688519 0.004154166 0.1313869
#> [22,] 0.6217129 0.56415821 0.55650754 0.521765004 0.339417104 0.1509063
#> [23,] 0.3255340 0.36927006 0.25188183 0.524599020 0.909987414 1.0342271
#> [24,] 0.6278218 0.61706443 0.53719272 0.430513668 0.313061081 0.1967150
#> [25,] 0.1246781 0.43106827 0.68688691 0.708727338 0.383067536 0.1488836
#>                V4        V40        V41        V42         V43         V44
#>  [1,] 0.065943673 0.51337464 0.47426559 0.42369370 0.483541305  0.56331180
#>  [2,] 0.062845577 0.13666940 0.18533781 0.18641281 0.106499693  0.09343454
#>  [3,] 0.159375871 0.56292107 0.41868498 0.45907483 0.450412677  0.39620622
#>  [4,] 0.064670302 0.23917076 0.56018318 0.60512696 0.489321032  0.44098783
#>  [5,] 0.079549546 0.24100657 0.21304643 0.24937548 0.228500721  0.14846500
#>  [6,] 0.058763836 0.28451061 0.29640444 0.33368429 0.171835472  0.05912620
#>  [7,] 0.110386953 0.35567593 0.39718669 0.48175545 0.456024932  0.40585074
#>  [8,] 0.033020115 0.20871975 0.17578649 0.16337709 0.158722865  0.17130129
#>  [9,] 0.072963736 0.50755766 0.32735252 0.20431320 0.192285418  0.20528421
#> [10,] 0.048241568 0.29462120 0.34115629 0.43996295 0.385510249  0.16774921
#> [11,] 0.041012245 0.44999709 0.31332856 0.13050062 0.237652197  0.20579391
#> [12,] 0.025130221 0.29637888 0.10624009 0.16768140 0.297480722  0.40473545
#> [13,] 0.035790815 0.45066050 0.27407772 0.44450244 0.449918872  0.48564552
#> [14,] 0.033900000 0.48160000 0.29170000 0.17690000 0.113600000  0.07010000
#> [15,] 0.026518178 0.09648298 0.05450931 0.02217804 0.067411807  0.10556836
#> [16,] 0.029529385 0.29292408 0.17634593 0.09011528 0.137384542  0.14838032
#> [17,] 0.017127033 0.57680675 0.38155849 0.33362516 0.240174174  0.13164139
#> [18,] 0.012041575 0.27956565 0.20787880 0.11014241 0.014479357 -0.03013713
#> [19,] 0.059596275 0.36107439 0.13940766 0.15895959 0.266134881  0.28111594
#> [20,] 0.023502248 0.15411868 0.18076022 0.17597218 0.159094953  0.11231641
#> [21,] 0.034764808 0.21220334 0.13684396 0.01677483 0.007605069  0.05894980
#> [22,] 0.013041161 0.09693821 0.08219309 0.18519226 0.266931466  0.30451234
#> [23,] 0.008958834 0.95511637 0.60790617 0.50286489 0.315299856  0.32451304
#> [24,] 0.025321596 0.17424962 0.18480014 0.13811984 0.158001393  0.10175008
#> [25,] 0.070185653 0.35351912 0.43520690 0.37300132 0.250211712  0.20302732
#>              V45        V46        V47        V48        V49          V5
#>  [1,] 0.49994908 0.31566010 0.19712166 0.17477787 0.10776506 0.101200726
#>  [2,] 0.06792811 0.14192956 0.14964787 0.11379588 0.05576210 0.094923167
#>  [3,] 0.51363131 0.41131974 0.07953804 0.12740003 0.06326495 0.135706432
#>  [4,] 0.55092648 0.56840076 0.39976974 0.23210690 0.13653071 0.069575017
#>  [5,] 0.10968017 0.07961872 0.07625463 0.06107568 0.03912307 0.098999604
#>  [6,] 0.11389009 0.19801738 0.22457555 0.15162283 0.05123912 0.100483923
#>  [7,] 0.52982728 0.55249284 0.31551482 0.24663937 0.12929271 0.241547078
#>  [8,] 0.14519413 0.14309263 0.12754619 0.08520607 0.04435245 0.068826879
#>  [9,] 0.39123868 0.28101577 0.14226606 0.16355861 0.10829101 0.055087327
#> [10,] 0.16117372 0.11877722 0.12813023 0.12783018 0.06480502 0.068288937
#> [11,] 0.26414099 0.24181731 0.20128605 0.13250891 0.08804840 0.033396551
#> [12,] 0.36433203 0.22100828 0.12057382 0.06052554 0.01765755 0.048726075
#> [13,] 0.48261264 0.20598180 0.14687332 0.18333054 0.10385799 0.085419927
#> [14,] 0.15780000 0.19380000 0.11060000 0.06930000 0.01760000 0.030500000
#> [15,] 0.09276806 0.03917011 0.04198095 0.03660625 0.02323882 0.063062459
#> [16,] 0.12583901 0.06555563 0.03760107 0.04610421 0.02417775 0.026929385
#> [17,] 0.06155101 0.01651275 0.06291452 0.06622193 0.02874883 0.025549972
#> [18,] 0.04558958 0.06552059 0.07043861 0.04329336 0.01120401 0.025759732
#> [19,] 0.19300599 0.07600277 0.06051072 0.04170788 0.02245355 0.117665081
#> [20,] 0.07934272 0.05540294 0.03513942 0.03858053 0.01987687 0.034053511
#> [21,] 0.08067171 0.03823324 0.06900641 0.04640716 0.02566775 0.048797091
#> [22,] 0.25739806 0.17910032 0.10252228 0.10348980 0.06134789 0.024067906
#> [23,] 0.45456546 0.37857752 0.27488465 0.15619122 0.06722222 0.003518773
#> [24,] 0.05573515 0.05298496 0.03947464 0.03543838 0.01292534 0.048098615
#> [25,] 0.22633681 0.19653359 0.18992129 0.17499556 0.11323441 0.072890821
#>               V50         V51          V52         V53         V54         V55
#>  [1,] 0.030465078 0.013693183 0.0168203273 0.010266980 0.009923642 0.001488424
#>  [2,] 0.011863783 0.019891510 0.0101075506 0.005968343 0.006412767 0.006099712
#>  [3,] 0.038407423 0.017964297 0.0354018965 0.034570789 0.024152255 0.016630932
#>  [4,] 0.052085511 0.053309494 0.0353830144 0.017777257 0.021597071 0.015136435
#>  [5,] 0.019673856 0.018778713 0.0156978740 0.015315393 0.015228806 0.015116493
#>  [6,] 0.011032969 0.021444381 0.0169361124 0.010566094 0.007790388 0.006644475
#>  [7,] 0.081212712 0.024237770 0.0205431332 0.035287161 0.023311606 0.043927942
#>  [8,] 0.017420438 0.017368596 0.0129100915 0.012089113 0.013431704 0.007221643
#>  [9,] 0.043810246 0.026834813 0.0106782409 0.010125327 0.031769314 0.018527331
#> [10,] 0.015459326 0.016946619 0.0117241933 0.012258210 0.011331765 0.005616388
#> [11,] 0.028638499 0.017484468 0.0071616528 0.012558421 0.016000952 0.014434652
#> [12,] 0.010506209 0.005353623 0.0076892949 0.009046213 0.011238851 0.003812574
#> [13,] 0.024065648 0.033673346 0.0176379055 0.006044697 0.009637464 0.010908334
#> [14,] 0.020500000 0.030900000 0.0212000000 0.009100000 0.005600000 0.008600000
#> [15,] 0.007753268 0.015925901 0.0107498136 0.007331479 0.005916074 0.002784304
#> [16,] 0.007331312 0.008924181 0.0081726825 0.012438374 0.004385877 0.005036376
#> [17,] 0.014238430 0.007150678 0.0135156999 0.021434816 0.013547344 0.014964128
#> [18,] 0.027501176 0.019136652 0.0148046517 0.026555507 0.028372494 0.010644043
#> [19,] 0.025127583 0.003603611 0.0199384170 0.006460123 0.018618133 0.011843488
#> [20,] 0.007945219 0.008144544 0.0103841817 0.005336199 0.008337809 0.010553960
#> [21,] 0.016160455 0.016951230 0.0002369658 0.006840258 0.012911833 0.012315981
#> [22,] 0.017029767 0.010996791 0.0080612330 0.002440907 0.008993581 0.005542605
#> [23,] 0.030006859 0.023807407 0.0125334710 0.002592180 0.014880795 0.008738409
#> [24,] 0.012107650 0.006701007 0.0071631255 0.007680112 0.007358789 0.006567040
#> [25,] 0.027126214 0.017144775 0.0145462031 0.011580470 0.011195211 0.006370795
#>               V56          V57          V58          V59          V6
#>  [1,] 0.007384506  0.007116824 0.0047244365 7.621524e-04 0.110570683
#>  [2,] 0.009701455  0.005614578 0.0050760135 5.226313e-03 0.121394470
#>  [3,] 0.023500222  0.013519662 0.0109041113 6.773354e-03 0.066042627
#>  [4,] 0.012506517  0.012202463 0.0157017365 1.467744e-02 0.102025508
#>  [5,] 0.013368750  0.008516868 0.0134713864 1.267511e-02 0.112365112
#>  [6,] 0.004722279  0.004375254 0.0047768077 2.911637e-03 0.138006684
#>  [7,] 0.038656381  0.034683330 0.0431517173 2.400486e-02 0.120805633
#>  [8,] 0.007438306  0.006941988 0.0069986174 5.947735e-03 0.107974450
#>  [9,] 0.023981078  0.014386510 0.0195587208 2.206758e-02 0.118160568
#> [10,] 0.004649080  0.005047795 0.0099529294 8.782448e-03 0.085458142
#> [11,] 0.010107429  0.005765356 0.0128025724 9.679414e-03 0.034868814
#> [12,] 0.007462617  0.003188787 0.0048272768 1.070911e-02 0.091290880
#> [13,] 0.004690418  0.004228035 0.0034390010 4.327039e-03 0.133940011
#> [14,] 0.009200000  0.007000000 0.0116000000 6.000000e-03 0.117200000
#> [15,] 0.006340460  0.004711306 0.0032120691 5.474657e-03 0.112276095
#> [16,] 0.006951568  0.006423182 0.0020221832 5.112125e-03 0.034912984
#> [17,] 0.005262842  0.004233626 0.0073158834 8.364883e-03 0.054948425
#> [18,] 0.012146810  0.003114164 0.0104125721 8.503847e-06 0.006449942
#> [19,] 0.002809412  0.003531886 0.0052492513 1.018118e-02 0.241111099
#> [20,] 0.005722360  0.005045814 0.0057661890 5.285452e-03 0.071182903
#> [21,] 0.005992318 -0.001074123 0.0001915042 2.740655e-03 0.079410715
#> [22,] 0.002534047  0.003378163 0.0037824418 6.950535e-03 0.092460912
#> [23,] 0.006801508  0.003439504 0.0033596695 8.356615e-04 0.020111366
#> [24,] 0.005296707  0.008335752 0.0068647962 6.699731e-03 0.081385569
#> [25,] 0.006533073  0.016223722 0.0106443459 7.590808e-03 0.085283248
#>               V60         V7         V8         V9
#>  [1,] 0.002519212 0.15887654 0.19640450 0.34106200
#>  [2,] 0.004168796 0.09011097 0.13914823 0.20505892
#>  [3,] 0.012848334 0.11320732 0.32413837 0.61867495
#>  [4,] 0.009636357 0.11045119 0.15544011 0.21612529
#>  [5,] 0.009149922 0.13109328 0.12047699 0.14148488
#>  [6,] 0.004151148 0.11914712 0.12049503 0.12496377
#>  [7,] 0.009685655 0.17356116 0.09983732 0.24260889
#>  [8,] 0.005395448 0.12725725 0.14247569 0.20416557
#>  [9,] 0.006628335 0.12549061 0.43288622 0.59615921
#> [10,] 0.002438022 0.11568670 0.13509803 0.19395971
#> [11,] 0.004872221 0.04748928 0.12286040 0.25823413
#> [12,] 0.005246499 0.15588435 0.18249108 0.37829236
#> [13,] 0.004070758 0.12769717 0.16868945 0.27888178
#> [14,] 0.011000000 0.14500000 0.06380000 0.07400000
#> [15,] 0.004089801 0.11724263 0.06460417 0.10349876
#> [16,] 0.004695935 0.02823359 0.03114537 0.04695421
#> [17,] 0.003361544 0.02904536 0.07389567 0.10469523
#> [18,] 0.003785050 0.02142807 0.03485716 0.08861099
#> [19,] 0.006502190 0.20131632 0.19169585 0.30329108
#> [20,] 0.004048011 0.08445083 0.09240107 0.09660019
#> [21,] 0.006757418 0.09708902 0.03195753 0.02470182
#> [22,] 0.003621837 0.10014002 0.05278830 0.09159956
#> [23,] 0.008210558 0.09767365 0.13310508 0.05277777
#> [24,] 0.005781410 0.07985585 0.08498976 0.10413943
#> [25,] 0.006374106 0.18492839 0.24418108 0.24908342
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478