Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12         V13        V14
#>  [1,] 0.030700000 0.27920000 0.38770000 0.49920000 0.498100000 0.49720000
#>  [2,] 0.048736787 0.22965992 0.32667157 0.45555928 0.547775865 0.55770429
#>  [3,] 0.031034165 0.43148674 0.41362903 0.19275481 0.494717347 0.97123987
#>  [4,] 0.019928466 0.65266934 0.69532284 0.48717314 0.375481680 0.28602479
#>  [5,] 0.043313755 0.51319619 0.34642769 0.09895625 0.238684868 0.33939954
#>  [6,] 0.018741803 0.08202122 0.22966633 0.41122810 0.382717809 0.51329758
#>  [7,] 0.036307950 0.20117960 0.27887012 0.28616605 0.280348973 0.26828428
#>  [8,] 0.037702134 0.25477315 0.25286054 0.27352620 0.298289643 0.29523590
#>  [9,] 0.079000000 0.66090000 0.50020000 0.25830000 0.165000000 0.43470000
#> [10,] 0.022873013 0.23468260 0.25723904 0.27014980 0.295695162 0.34807397
#> [11,] 0.031716619 0.24197957 0.31457363 0.33326419 0.273748015 0.12690442
#> [12,] 0.060487746 0.10753301 0.25820954 0.37073304 0.462337952 0.46103280
#> [13,] 0.024581519 0.30424907 0.37395061 0.40615787 0.384635764 0.30356408
#> [14,] 0.025204047 0.08185892 0.12479816 0.27846113 0.363191082 0.33618566
#> [15,] 0.025280488 0.36137655 0.28758133 0.10523372 0.251253419 0.51780871
#> [16,] 0.000670112 0.09229098 0.11248995 0.04690392 0.008160651 0.03666197
#> [17,] 0.018353344 0.06415952 0.12431192 0.17692716 0.254995179 0.28562303
#> [18,] 0.014095896 0.13248081 0.10265296 0.06584393 0.068814777 0.09159564
#> [19,] 0.020835942 0.11347956 0.10020010 0.11052671 0.126640527 0.14822492
#> [20,] 0.028822124 0.05306577 0.05775986 0.12818678 0.186089657 0.19372400
#> [21,] 0.010062952 0.10114752 0.17507983 0.14479996 0.128346266 0.20985376
#> [22,] 0.035609504 0.03256025 0.12294952 0.20342742 0.226734431 0.07148671
#> [23,] 0.013796375 0.11619669 0.11851172 0.10428291 0.167653333 0.21812641
#> [24,] 0.009300000 0.13600000 0.21320000 0.37380000 0.373800000 0.26730000
#> [25,] 0.022401190 0.04977523 0.08095523 0.17469752 0.136134599 0.17805414
#>              V15        V16        V17       V18        V19           V2
#>  [1,] 0.56070000 0.73390000 0.82300000 0.9173000 0.99750000  0.052300000
#>  [2,] 0.64599702 0.75709228 0.91879519 0.9226264 0.88890197  0.091558666
#>  [3,] 0.92220751 0.86053427 0.71260387 0.7839768 0.87796023  0.037459361
#>  [4,] 0.19124005 0.09721108 0.16138796 0.3722752 0.61088923  0.026829310
#>  [5,] 0.38619980 0.39941185 0.17808943 0.1653540 0.28484569  0.075301350
#>  [6,] 0.65800914 0.66066278 0.79583659 0.8864179 0.97739341  0.016092475
#>  [7,] 0.21464962 0.20201964 0.28256531 0.3438856 0.39733770  0.057070698
#>  [8,] 0.24003287 0.29571786 0.32937186 0.2529901 0.30364485  0.050561485
#>  [9,] 0.45150000 0.45790000 0.33660000 0.4000000 0.53250000  0.070700000
#> [10,] 0.36806470 0.42995686 0.49075445 0.5863649 0.68039155  0.010418651
#> [11,] 0.13098899 0.17111340 0.19989227 0.2492322 0.23705634  0.017662396
#> [12,] 0.57598780 0.56803092 0.73166027 0.8026105 0.93538154  0.072655625
#> [13,] 0.23087367 0.24847352 0.25971400 0.3127194 0.50282187  0.031816776
#> [14,] 0.14965942 0.14574891 0.21406404 0.4163948 0.47956761  0.038654799
#> [15,] 0.65406040 0.84050045 0.89883282 0.8290024 0.57488572  0.042078714
#> [16,] 0.10209890 0.07077326 0.07800262 0.1422696 0.27810086 -0.007810152
#> [17,] 0.30172224 0.31041236 0.29506422 0.3516416 0.46418590  0.020378624
#> [18,] 0.09653066 0.11087982 0.19533142 0.2565359 0.33260608  0.010862664
#> [19,] 0.15335099 0.17313573 0.14609107 0.1536606 0.12915867  0.017383100
#> [20,] 0.16934534 0.12322152 0.13360174 0.0900607 0.09071668  0.054188561
#> [21,] 0.28973020 0.28606450 0.22148616 0.1880707 0.21718083  0.021794545
#> [22,] 0.04456019 0.16738859 0.19630689 0.3246497 0.17931377  0.032985698
#> [23,] 0.22156058 0.22432459 0.17440579 0.2129218 0.26563670  0.016096183
#> [24,] 0.23330000 0.53670000 0.73120000 0.7659000 0.62710000  0.026900000
#> [25,] 0.20843506 0.37116631 0.71950227 0.5981585 0.25837967  0.027927849
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.9911000 0.8240000 0.6498000 0.5980000 0.4862000 0.3150000 0.1543000
#>  [2,] 0.7953511 0.7346308 0.5900063 0.4272821 0.3676750 0.2447503 0.1545651
#>  [3,] 0.8797317 0.8999877 0.9284140 0.9953223 0.8920273 0.7425851 0.7751609
#>  [4,] 0.7355019 0.8122511 0.8765697 0.8827097 0.9592725 0.8840934 0.8192497
#>  [5,] 0.8903232 0.9810163 0.3649017 0.1532736 0.1936301 0.1198078 0.3373042
#>  [6,] 1.0131124 0.9008587 0.8874436 0.9510640 0.8913173 0.5410266 0.3706854
#>  [7,] 0.3984767 0.4157267 0.4499045 0.5386092 0.6152376 0.6378090 0.7114262
#>  [8,] 0.3773935 0.4199174 0.4728837 0.5421708 0.6148097 0.6260330 0.6633201
#>  [9,] 0.9010000 0.9939000 0.3689000 0.1012000 0.0248000 0.2318000 0.3981000
#> [10,] 0.7551519 0.8178200 0.8975955 0.9597987 0.9450019 0.9108099 0.8823994
#> [11,] 0.2788678 0.3855753 0.6184514 0.7949957 0.8601637 0.9596934 0.9992700
#> [12,] 0.9238046 0.9831853 1.0043928 0.9535634 0.8982600 0.7360430 0.5708020
#> [13,] 0.6616620 0.7943501 0.8293080 0.7971100 0.8478049 0.8582395 0.9078352
#> [14,] 0.7407232 0.7386899 0.6246481 0.5834846 0.5963352 0.7529491 0.7927402
#> [15,] 0.4005036 0.4833499 0.5998303 0.7621317 0.8856422 0.8155668 0.7212748
#> [16,] 0.3221527 0.3182658 0.4870126 0.7758604 0.9310379 0.9328325 0.7880558
#> [17,] 0.5416689 0.5602131 0.5829954 0.6352033 0.7248954 0.8204257 0.8450199
#> [18,] 0.4557986 0.5677094 0.7617074 0.8762417 0.8420050 0.7739510 0.8261425
#> [19,] 0.1362382 0.2430746 0.2902723 0.4379034 0.6508415 0.9332210 1.0303866
#> [20,] 0.1296631 0.1846677 0.2027317 0.1300284 0.1795279 0.2854567 0.4775147
#> [21,] 0.2943109 0.2730322 0.4257634 0.6434483 0.6792066 0.7630287 0.8109944
#> [22,] 0.2716985 0.5988527 0.6868141 0.7932665 0.8434945 0.7536663 0.6404646
#> [23,] 0.2824639 0.2753617 0.3277448 0.4226362 0.5545779 0.5762886 0.6406730
#> [24,] 0.4395000 0.4330000 0.4326000 0.5544000 0.7360000 0.8589000 0.8989000
#> [25,] 0.1989368 0.3756964 0.5991670 0.6834298 0.7146961 0.8327003 0.9182441
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.0989000 0.0284000 0.1008000  0.06530000 0.2636000 0.2694000 0.2930000
#>  [2,] 0.1718937 0.1789601 0.1796280  0.09174249 0.1635803 0.2528559 0.3075306
#>  [3,] 0.6391896 0.4817682 0.4687899  0.01930427 0.3338830 0.2710976 0.1076844
#>  [4,] 0.8336826 0.7920004 0.5664311  0.04528511 0.3535490 0.3151698 0.3428037
#>  [5,] 0.2673317 0.4895690 0.4451785  0.10087876 0.6862192 0.2196941 0.2970531
#>  [6,] 0.2675749 0.3308544 0.3370310  0.04998286 0.2885548 0.2321896 0.3306034
#>  [7,] 0.8156923 0.8250650 0.7509364  0.07130973 0.5905925 0.5004110 0.4198376
#>  [8,] 0.7666739 0.8099693 0.7886404  0.06343920 0.8555273 0.8299561 0.7890410
#>  [9,] 0.2259000 0.5247000 0.6898000  0.03520000 0.8316000 0.4326000 0.3741000
#> [10,] 0.9220592 0.9187368 0.7839507  0.02526872 0.5565893 0.3508304 0.3218052
#> [11,] 0.9690105 0.9660968 0.8738486  0.02827250 0.7474457 0.7389302 0.6214511
#> [12,] 0.4228361 0.3071244 0.2603801  0.12360209 0.3489309 0.2949942 0.1452026
#> [13,] 0.9461953 0.9208470 0.7920662  0.03943866 0.5735424 0.4165849 0.3654250
#> [14,] 0.7242410 0.5286295 0.4337379  0.07176089 0.3209926 0.3316033 0.4091526
#> [15,] 0.7443874 0.5948688 0.3389095  0.06225632 0.3890127 0.5284333 0.4492854
#> [16,] 0.7248058 0.7491099 0.6092181 -0.02178574 0.3987767 0.2241022 0.2527972
#> [17,] 0.8873790 0.8726842 0.7612408  0.02506140 0.6397258 0.5541843 0.4296860
#> [18,] 0.8584681 0.8690997 0.8856090  0.01556617 0.8036508 0.7229593 0.6570664
#> [19,] 0.8227380 0.6170500 0.5296764  0.01092806 0.3647168 0.2688338 0.0870121
#> [20,] 0.5662886 0.6102900 0.5965177  0.02444227 0.6207389 0.7328261 0.7557809
#> [21,] 0.6982476 0.7609623 0.7251820  0.02451952 0.5996689 0.5113149 0.4181796
#> [22,] 0.5744057 0.4072289 0.4389547  0.02219182 0.5149244 0.5581154 0.5408861
#> [23,] 0.6899622 0.8384407 0.9328485  0.02059886 0.8710795 0.7553405 0.6104303
#> [24,] 0.9420000 0.9401000 0.9379000  0.02170000 0.8575000 0.7284000 0.6700000
#> [25,] 0.8937295 0.8021978 0.8452171  0.03859097 0.9025026 0.8784272 0.7420755
#>              V33        V34       V35       V36        V37       V38       V39
#>  [1,] 0.29250000 0.39980000 0.3660000 0.3172000 0.46090000 0.4374000 0.1820000
#>  [2,] 0.32842663 0.39863371 0.4463199 0.2659715 0.19553025 0.2911877 0.3127941
#>  [3,] 0.04301805 0.13008037 0.1718385 0.2093940 0.13142706 0.1819456 0.3227710
#>  [4,] 0.57385032 0.65988033 0.5640885 0.4862359 0.45902523 0.3357579 0.3658042
#>  [5,] 0.58367250 0.78977913 0.8348326 0.8924559 0.80910961 0.9144721 0.7157937
#>  [6,] 0.38720949 0.22804486 0.2070748 0.2285969 0.34944462 0.5833169 0.4552304
#>  [7,] 0.38500737 0.28676827 0.1932359 0.1963456 0.24970062 0.1988756 0.2411347
#>  [8,] 0.60539441 0.53169062 0.5319408 0.4305244 0.36232613 0.4522106 0.4584842
#>  [9,] 0.57560000 0.80430000 0.7963000 0.7174000 0.70560000 0.8148000 0.7601000
#> [10,] 0.17966379 0.02082107 0.1153060 0.2467678 0.27760892 0.3104153 0.3436751
#> [11,] 0.45475995 0.21137156 0.1052887 0.1080110 0.12884957 0.1433225 0.1701136
#> [12,] 0.29594552 0.50000849 0.6048935 0.5335697 0.45442548 0.4641566 0.3944950
#> [13,] 0.30962334 0.24669097 0.1502706 0.1667836 0.18066362 0.2382843 0.3155918
#> [14,] 0.49763979 0.45765523 0.4398884 0.5134685 0.55428253 0.3387128 0.2704583
#> [15,] 0.25074977 0.28695086 0.6599247 0.6891329 0.54906601 0.4368443 0.3337853
#> [16,] 0.25899496 0.29611311 0.1903260 0.2280655 0.33036014 0.2867506 0.3010088
#> [17,] 0.32329078 0.18033900 0.1746222 0.2935438 0.29422352 0.2512022 0.2440180
#> [18,] 0.64557208 0.67439296 0.6440161 0.5390514 0.43499705 0.3311467 0.2361087
#> [19,] 0.11684286 0.14591804 0.1939574 0.1785511 0.03985005 0.1044038 0.1685443
#> [20,] 0.77407507 0.80072646 0.9413081 1.0059435 0.88416843 0.7356178 0.6076845
#> [21,] 0.51483440 0.78085763 0.8315217 0.6895597 0.37938639 0.2596228 0.2786920
#> [22,] 0.62693192 0.59448242 0.5934616 0.5720653 0.53421713 0.4529499 0.4783186
#> [23,] 0.57775141 0.54489317 0.4780178 0.4212981 0.30698643 0.2188440 0.1921252
#> [24,] 0.75470000 0.87730000 0.9919000 0.9922000 0.94190000 0.8388000 0.6605000
#> [25,] 0.67985365 0.74528524 0.8703271 0.9486464 0.86305162 0.7057605 0.5405305
#>                V4        V40        V41       V42       V43        V44
#>  [1,]  0.05210000 0.33760000 0.62020000 0.4448000 0.1863000 0.14200000
#>  [2,]  0.12262956 0.17076488 0.09042402 0.0997778 0.1185489 0.05801071
#>  [3,]  0.06974191 0.33678521 0.23569063 0.1138151 0.2439903 0.25669203
#>  [4,]  0.05924685 0.36785085 0.24574682 0.2240129 0.2499373 0.21861616
#>  [5,]  0.09465337 0.48883918 0.43425422 0.2896501 0.1545881 0.18133493
#>  [6,]  0.14459170 0.25917930 0.12451175 0.1886961 0.3174106 0.14216981
#>  [7,]  0.10466290 0.29250051 0.29998438 0.3511696 0.2617465 0.15002329
#>  [8,]  0.07117360 0.35996201 0.51394614 0.5127992 0.4109572 0.47031970
#>  [9,]  0.16600000 0.60340000 0.45540000 0.4729000 0.4478000 0.37220000
#> [10,]  0.06646968 0.35741906 0.36373275 0.3818616 0.3046927 0.22930031
#> [11,]  0.04968247 0.06086152 0.13493000 0.1116220 0.1517397 0.16635177
#> [12,]  0.15376048 0.23922367 0.23760654 0.4001745 0.3701119 0.23547356
#> [13,]  0.03950763 0.27892713 0.13051870 0.1937761 0.2767425 0.30424342
#> [14,]  0.05460591 0.44280059 0.37397612 0.4321946 0.3737741 0.24916070
#> [15,]  0.04894830 0.34496917 0.24816164 0.2402405 0.2542889 0.22809631
#> [16,] -0.03716585 0.25724200 0.20548512 0.1634444 0.1425952 0.19105792
#> [17,]  0.02998609 0.15586257 0.16529972 0.2208166 0.1728644 0.12396784
#> [18,]  0.03124627 0.22478143 0.19948280 0.1159282 0.1139324 0.07529810
#> [19,]  0.01816301 0.24185633 0.31622203 0.2915379 0.1868060 0.06327931
#> [20,]  0.02009179 0.48987435 0.38419711 0.3420298 0.3016510 0.24517265
#> [21,]  0.05057352 0.32613111 0.27529232 0.3189278 0.3522310 0.25015981
#> [22,]  0.00637409 0.64784426 0.73607428 0.7179604 0.7429936 0.54053738
#> [23,]  0.02022710 0.17044113 0.15667865 0.1820030 0.2364965 0.20689013
#> [24,]  0.03390000 0.48160000 0.29170000 0.1769000 0.1136000 0.07010000
#> [25,]  0.02553956 0.34843681 0.27076244 0.2157764 0.2681260 0.19379537
#>              V45         V46         V47        V48        V49          V5
#>  [1,] 0.05890000  0.05760000  0.06720000 0.02690000 0.02450000  0.06110000
#>  [2,] 0.07056812  0.06702956  0.08037910 0.08838895 0.06219254  0.14275411
#>  [3,] 0.13886745  0.09431309  0.12490376 0.12369096 0.10270062  0.10290176
#>  [4,] 0.15559760  0.08976302  0.09936038 0.12361216 0.05257634  0.03517304
#>  [5,] 0.34025660  0.24784296  0.17762247 0.22581114 0.11112069  0.03988679
#>  [6,] 0.03243607 -0.01202115 -0.04117908 0.02080280 0.00187744  0.23711403
#>  [7,] 0.15688923  0.16469619  0.16755778 0.13959856 0.07137932  0.11838851
#>  [8,] 0.51530789  0.43670413  0.29276968 0.19636533 0.10720287  0.08614618
#>  [9,] 0.46930000  0.38390000  0.07680000 0.14670000 0.07770000  0.13300000
#> [10,] 0.13737717  0.14253479  0.14784524 0.09011601 0.05496969  0.12393801
#> [11,] 0.10483706  0.04489357  0.09549932 0.11326990 0.06338600  0.06043390
#> [12,] 0.17395715  0.11340866  0.07348916 0.02224510 0.04202726  0.19225566
#> [13,] 0.28982389  0.20408531  0.14153004 0.10355414 0.06030582  0.06887675
#> [14,] 0.13761466  0.08217397  0.04467124 0.05294408 0.03103702  0.06299959
#> [15,] 0.20131031  0.11895524  0.10756173 0.09300176 0.05513308  0.08188299
#> [16,] 0.15591563  0.11367156  0.02648971 0.02983967 0.02385257 -0.02046518
#> [17,] 0.11258187  0.10388359  0.09420560 0.06715083 0.03010612  0.03456116
#> [18,] 0.03736986  0.03023647  0.03115430 0.03458515 0.01377157  0.05127589
#> [19,] 0.06910448  0.08668180  0.04204616 0.02033853 0.02456939  0.06094865
#> [20,] 0.20895613  0.13570332  0.10379527 0.08815837 0.04627929  0.02775912
#> [21,] 0.10570656  0.05604122  0.07373079 0.06559449 0.03079849  0.08128363
#> [22,] 0.40760561  0.29583740  0.17488722 0.10254846 0.02904736  0.01492557
#> [23,] 0.12771198  0.09549476  0.06638303 0.04591805 0.02617510  0.03427380
#> [24,] 0.15780000  0.19380000  0.11060000 0.06930000 0.01760000  0.03050000
#> [25,] 0.20141796  0.17575059  0.10157695 0.02132531 0.01258470  0.01012021
#>               V50          V51          V52         V53         V54
#>  [1,] 0.019000000 0.0063000000 0.0321000000 0.018900000 0.013700000
#>  [2,] 0.026807817 0.0208186900 0.0277666912 0.012291371 0.012363758
#>  [3,] 0.028232144 0.0312454430 0.0078691769 0.015612223 0.008756213
#>  [4,] 0.022216874 0.0117273332 0.0126776692 0.008698963 0.020803206
#>  [5,] 0.030566345 0.0389722916 0.0183606854 0.013925059 0.029081886
#>  [6,] 0.001639972 0.0029414685 0.0234517055 0.013012390 0.009628517
#>  [7,] 0.013424799 0.0184341711 0.0157120611 0.011433426 0.009544920
#>  [8,] 0.042093223 0.0329377347 0.0244052741 0.014748568 0.018997798
#>  [9,] 0.046900000 0.0193000000 0.0298000000 0.039000000 0.029400000
#> [10,] 0.023309347 0.0165583811 0.0181148360 0.017666922 0.013287018
#> [11,] 0.018404405 0.0079177192 0.0078124546 0.008137979 0.011938434
#> [12,] 0.008900637 0.0152947821 0.0121796023 0.008723315 0.025538618
#> [13,] 0.017202807 0.0210808911 0.0123753439 0.007462844 0.008134478
#> [14,] 0.016244082 0.0045761947 0.0106728637 0.012634134 0.019781316
#> [15,] 0.023863574 0.0119105737 0.0162527773 0.006444947 0.014600372
#> [16,] 0.002989717 0.0005033437 0.0071729935 0.006612861 0.006885553
#> [17,] 0.018894369 0.0092494322 0.0077440989 0.009287728 0.009083503
#> [18,] 0.010889365 0.0059222467 0.0062881909 0.007827122 0.005359978
#> [19,] 0.009202193 0.0143029414 0.0067992518 0.005688983 0.007694367
#> [20,] 0.009649185 0.0110238369 0.0105631374 0.019302976 0.007723265
#> [21,] 0.012244656 0.0255736530 0.0092618956 0.011017825 0.011407452
#> [22,] 0.023089043 0.0075191450 0.0009599579 0.006063958 0.002078234
#> [23,] 0.014824665 0.0126910690 0.0118751305 0.010891642 0.006994252
#> [24,] 0.020500000 0.0309000000 0.0212000000 0.009100000 0.005600000
#> [25,] 0.036033720 0.0041361858 0.0078643615 0.002707321 0.013449447
#>                V55          V56         V57          V58         V59
#>  [1,]  0.027700000 0.0152000000 0.005200000 0.0121000000 0.012400000
#>  [2,]  0.017078171 0.0102446593 0.003833466 0.0103794793 0.018458101
#>  [3,]  0.035839951 0.0089520694 0.012167418 0.0268980323 0.012840286
#>  [4,]  0.012660833 0.0116103065 0.007198017 0.0163608104 0.015213448
#>  [5,]  0.016419443 0.0265883339 0.018805450 0.0167234788 0.020408220
#>  [6,]  0.030039480 0.0354649571 0.009358766 0.0050049281 0.020236699
#>  [7,]  0.008058774 0.0070773732 0.006556853 0.0089177575 0.009015918
#>  [8,]  0.019719863 0.0152011793 0.012366388 0.0135675085 0.015315962
#>  [9,]  0.017500000 0.0249000000 0.014100000 0.0073000000 0.002500000
#> [10,]  0.003762542 0.0105146491 0.010741991 0.0115478024 0.010849383
#> [11,]  0.005468151 0.0061737761 0.004242347 0.0048631631 0.007096510
#> [12,]  0.010679781 0.0146945078 0.025694204 0.0316547124 0.020340829
#> [13,]  0.007234134 0.0045708659 0.004479033 0.0047515953 0.004813381
#> [14,]  0.014776377 0.0135684350 0.013600701 0.0102127085 0.007716051
#> [15,]  0.009983022 0.0034666011 0.009227454 0.0067793252 0.011982282
#> [16,] -0.001354235 0.0004180108 0.001413671 0.0058442948 0.002227090
#> [17,]  0.007735470 0.0081627251 0.008563983 0.0079351731 0.003542982
#> [18,]  0.004269947 0.0048543870 0.007416858 0.0052083401 0.004265831
#> [19,]  0.007657676 0.0041638084 0.001701196 0.0008474071 0.004176021
#> [20,]  0.016500927 0.0102734319 0.005176014 0.0075656704 0.007723013
#> [21,]  0.005518957 0.0103482516 0.011126706 0.0042703344 0.009712663
#> [22,]  0.001734137 0.0102040929 0.006758685 0.0068803750 0.009027920
#> [23,]  0.005808748 0.0036635994 0.005750948 0.0037386814 0.007225353
#> [24,]  0.008600000 0.0092000000 0.007000000 0.0116000000 0.006000000
#> [25,]  0.010517471 0.0020327096 0.011165913 0.0063184330 0.007445157
#>                V6         V60         V7         V8          V9
#>  [1,]  0.05770000 0.005500000 0.06650000 0.06640000 0.146000000
#>  [2,]  0.08840466 0.010166355 0.10847702 0.09731313 0.129890498
#>  [3,]  0.17459050 0.004004697 0.29187943 0.33287132 0.339913904
#>  [4,]  0.04080648 0.007741526 0.16176566 0.32673354 0.511242784
#>  [5,]  0.12322602 0.012405954 0.22576208 0.41304266 0.601182083
#>  [6,]  0.08303061 0.007585108 0.05204187 0.12755569 0.065140714
#>  [7,]  0.10470924 0.005313190 0.10285939 0.11000429 0.171027928
#>  [8,]  0.10649541 0.009199319 0.12331838 0.13082338 0.251489147
#>  [9,]  0.02260000 0.010100000 0.07710000 0.26780000 0.566400000
#> [10,]  0.12348108 0.007812139 0.13949607 0.16217422 0.194324213
#> [11,]  0.09979613 0.007597808 0.11412140 0.12387515 0.147845947
#> [12,]  0.18343808 0.007975281 0.16385826 0.09225829 0.084591900
#> [13,]  0.09380308 0.004578699 0.11136562 0.14230804 0.219669915
#> [14,]  0.11209255 0.010851041 0.11214661 0.10152911 0.038614567
#> [15,]  0.17358469 0.008961688 0.18577050 0.21208250 0.294094578
#> [16,] -0.00224896 0.002640502 0.04696050 0.11028358 0.064750073
#> [17,]  0.04595076 0.003909880 0.07663178 0.06260222 0.088320157
#> [18,]  0.07416853 0.004261929 0.07126359 0.09478915 0.114711809
#> [19,]  0.05997527 0.001925174 0.05817572 0.03929875 0.097785792
#> [20,]  0.07381983 0.004408097 0.07887201 0.05903747 0.001195628
#> [21,]  0.10596516 0.008735815 0.11206201 0.10145473 0.083606642
#> [22,]  0.06180472 0.003862317 0.07471385 0.12676714 0.074405191
#> [23,]  0.07795007 0.005639508 0.07581858 0.07988410 0.120039138
#> [24,]  0.11720000 0.011000000 0.14500000 0.06380000 0.074000000
#> [25,]  0.04338126 0.008645667 0.12412907 0.12991118 0.048751851
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841