Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1         V10        V11         V12        V13       V14
#>  [1,] 0.0356252544 0.148955580 0.20688428 0.319226640 0.32856535 0.3982871
#>  [2,] 0.0229262229 0.314155391 0.26898317 0.287018543 0.25144629 0.2063600
#>  [3,] 0.0484259380 0.185934794 0.17404880 0.244883302 0.29310804 0.2955684
#>  [4,] 0.0237746566 0.372226567 0.44762924 0.467841092 0.49553461 0.4021956
#>  [5,] 0.0914410025 0.609869936 0.45672145 0.214587044 0.14258308 0.3591415
#>  [6,] 0.0113857734 0.200566633 0.10329336 0.008468631 0.10681393 0.2599526
#>  [7,] 0.0254816866 0.370070805 0.38152681 0.207401181 0.46086715 0.8853735
#>  [8,] 0.0004745069 0.170655692 0.26893975 0.333028404 0.40886530 0.4830076
#>  [9,] 0.0249048351 0.221714986 0.27348727 0.295337879 0.25360827 0.1802976
#> [10,] 0.0527427486 0.342154911 0.54151070 0.541276674 0.54635914 0.5351738
#> [11,] 0.0343219938 0.079434043 0.19908682 0.299034770 0.31041350 0.2835521
#> [12,] 0.0444638313 0.114172624 0.08959155 0.122126865 0.26815254 0.3029559
#> [13,] 0.0269320297 0.277169797 0.33497156 0.394341101 0.40181571 0.4146099
#> [14,] 0.0132353940 0.123561833 0.06917636 0.020199710 0.03483685 0.1200255
#> [15,] 0.0132714433 0.072220268 0.06748048 0.172165399 0.08058106 0.1333735
#> [16,] 0.0148835082 0.007106058 0.13229071 0.356045010 0.31622785 0.2872576
#> [17,] 0.0067940280 0.060486899 0.07595556 0.092336015 0.20772851 0.1996090
#> [18,] 0.0269485596 0.092266008 0.24281514 0.365793580 0.29106700 0.2804719
#> [19,] 0.0327266174 0.160280603 0.25013571 0.297092656 0.37191637 0.3700563
#> [20,] 0.0037022656 0.187765543 0.11647219 0.091361105 0.18308513 0.1705551
#> [21,] 0.0176615102 0.103570907 0.10382754 0.098757063 0.11819350 0.1473707
#> [22,] 0.0219149612 0.151454644 0.13885833 0.079609559 0.05341178 0.1529197
#> [23,] 0.0106710756 0.119805727 0.09395258 0.102623539 0.12291839 0.1372044
#> [24,] 0.0455527807 0.292610031 0.25378524 0.091573225 0.17264739 0.4666186
#> [25,] 0.0155629471 0.115115716 0.15736968 0.152713012 0.18908408 0.2605091
#>              V15        V16        V17       V18       V19            V2
#>  [1,] 0.52080304 0.58224618 0.69315975 0.8312466 0.9469550  7.209354e-02
#>  [2,] 0.11747026 0.16701858 0.21037001 0.2621075 0.2035240  3.375993e-02
#>  [3,] 0.38234668 0.51926554 0.48219754 0.2321215 0.4500699  1.163659e-01
#>  [4,] 0.28753004 0.32613535 0.34556141 0.3257456 0.4789616  3.586615e-02
#>  [5,] 0.43905657 0.45221928 0.28787729 0.2999672 0.4338151  8.694261e-02
#>  [6,] 0.42578166 0.60441995 0.58647271 0.5823097 0.6464649  8.331368e-03
#>  [7,] 0.88001766 0.82864464 0.72975020 0.8144885 0.9077507  3.512453e-02
#>  [8,] 0.52232608 0.66139309 0.68893058 0.7969444 0.8944868  1.536355e-02
#>  [9,] 0.16301521 0.17895694 0.20349185 0.2605599 0.3318485  3.656479e-02
#> [10,] 0.61168124 0.79329461 0.94075350 0.9550605 0.8591196  7.938573e-02
#> [11,] 0.29571902 0.39872037 0.52236145 0.6028473 0.6640625  4.271081e-02
#> [12,] 0.34323464 0.36275727 0.38141550 0.2251380 0.2430978  7.071120e-02
#> [13,] 0.43038031 0.46408451 0.57247244 0.6368310 0.6703323  3.362338e-02
#> [14,] 0.07921557 0.05784004 0.19037209 0.2818653 0.2968887  1.482889e-02
#> [15,] 0.10765948 0.15311123 0.09939722 0.1674526 0.2359791  2.404400e-02
#> [16,] 0.01648140 0.02994059 0.19006916 0.4509645 0.3421714  7.735988e-03
#> [17,] 0.23088213 0.28715856 0.30940077 0.3198744 0.4103301  9.578252e-03
#> [18,] 0.31112584 0.44785457 0.58708444 0.4986750 0.4615479  5.668733e-02
#> [19,] 0.48453562 0.56250441 0.45767253 0.5024832 0.6347969  3.943834e-02
#> [20,] 0.22470503 0.30290730 0.33032237 0.2807940 0.3574156 -1.117523e-02
#> [21,] 0.17071348 0.18125622 0.15518527 0.1549632 0.1130315  1.558802e-02
#> [22,] 0.09794516 0.09517385 0.23678288 0.2468844 0.3018225 -1.124573e-05
#> [23,] 0.12959145 0.16678951 0.22787806 0.2718636 0.3476497  1.845850e-02
#> [24,] 0.63168469 0.71528809 0.76915045 0.6600200 0.4494391  6.925497e-02
#> [25,] 0.17336380 0.20517181 0.28116094 0.3763251 0.4161240  4.241264e-03
#>             V20        V21       V22        V23        V24       V25       V26
#>  [1,] 0.9821688 0.99441855 1.0795371 0.95093284 0.82565228 0.6251026 0.5004286
#>  [2,] 0.1217707 0.06346709 0.1948069 0.41423694 0.57603603 0.7060836 0.8420777
#>  [3,] 0.4539990 0.59055707 0.5486059 0.46697820 0.61482724 0.5483895 0.5224491
#>  [4,] 0.6910208 0.76847341 0.7778110 0.72378089 0.72735651 0.7674020 0.8960968
#>  [5,] 0.8829778 0.98607072 0.4011178 0.09535403 0.09022207 0.2213614 0.4036054
#>  [6,] 0.7443063 0.93010059 0.9906472 0.89941961 0.59945788 0.3789962 0.3914346
#>  [7,] 0.8634563 0.89116697 0.9443566 0.97561987 0.86864926 0.7503886 0.7764811
#>  [8,] 1.0229444 1.07929962 1.0697050 0.96951747 0.79999356 0.6042292 0.6742376
#>  [9,] 0.3604850 0.46466790 0.5987898 0.66721092 0.75284939 0.8151009 0.8669257
#> [10,] 0.8230424 0.76417552 0.4273041 0.09286131 0.11191741 0.2995707 0.2410133
#> [11,] 0.7215906 0.74356320 0.8276855 0.97799164 1.08746609 0.9883627 0.7307840
#> [12,] 0.3052288 0.31288417 0.4291888 0.50505531 0.40241799 0.4299306 0.4511125
#> [13,] 0.7335110 0.79971868 0.8478139 0.91284089 0.99347293 0.9952325 0.9898966
#> [14,] 0.4371264 0.57722827 0.6944278 0.81636893 0.83841981 0.9053371 0.9780743
#> [15,] 0.2584193 0.22566020 0.1784666 0.29825764 0.25797302 0.1136023 0.3252438
#> [16,] 0.5799879 0.67547104 0.7223538 0.79582622 0.85238581 1.0240856 0.8979391
#> [17,] 0.4580542 0.51474354 0.5752275 0.61182195 0.75028951 0.8790953 0.8714368
#> [18,] 0.6149525 0.39873597 0.2314064 0.38187457 0.89139342 0.9262228 0.5227705
#> [19,] 0.7240437 0.57371142 0.4212568 0.56182661 0.72666414 0.6517024 0.6096005
#> [20,] 0.3403316 0.37158236 0.4556947 0.50327521 0.51142038 0.6812259 0.8774038
#> [21,] 0.1286702 0.21876534 0.2676047 0.40453208 0.61759600 0.8727901 0.9939907
#> [22,] 0.4772843 0.60340152 0.7474249 0.90634440 0.92398998 0.9182758 0.9843490
#> [23,] 0.4176252 0.47963907 0.5959017 0.69364550 0.72317627 0.6763647 0.7134729
#> [24,] 0.3605841 0.56021095 0.7525323 0.83863161 0.87613617 0.8145171 0.6688884
#> [25,] 0.4415556 0.52137906 0.5789014 0.56197442 0.68123330 0.9826489 1.0111651
#>             V27       V28        V29           V3       V30       V31       V32
#>  [1,] 0.5550874 0.4628718 0.30621683 0.0927093152 0.3103450 0.2020257 0.2982020
#>  [2,] 0.9475565 0.9550471 0.91540944 0.0377841184 0.9559156 0.9286850 0.8249656
#>  [3,] 0.7322363 0.8764098 0.87405135 0.0745555812 0.9222555 0.8374380 0.6513586
#>  [4,] 0.9563197 0.8863039 0.72478514 0.0456370569 0.5384631 0.4633770 0.5565330
#>  [5,] 0.2413711 0.4706364 0.59401140 0.0535361780 0.7642822 0.3965088 0.3840633
#>  [6,] 0.5210594 0.7379523 0.72120738 0.0156169374 0.5591485 0.3030761 0.3795161
#>  [7,] 0.6047441 0.4225894 0.47597678 0.0261060412 0.3510679 0.2838078 0.1058249
#>  [8,] 0.7359938 0.7008838 0.54320319 0.0140018362 0.4363769 0.3473758 0.2291896
#>  [9,] 0.9326022 0.9514921 0.83019464 0.0466638593 0.6263408 0.5794372 0.4826296
#> [10,] 0.1979483 0.1959943 0.05748791 0.1299149341 0.1754630 0.1973209 0.2308624
#> [11,] 0.4431712 0.3141249 0.27272352 0.0583037169 0.3157209 0.4464357 0.4951181
#> [12,] 0.4504839 0.7012550 0.76996246 0.0924385214 0.7074088 0.7457936 0.7217843
#> [13,] 0.9628321 0.9486678 0.85580907 0.0354266661 0.6647469 0.4591034 0.3256726
#> [14,] 1.0012887 0.8671076 0.74611524 0.0206354468 0.6660456 0.6142029 0.5375630
#> [15,] 0.4829890 0.5083562 0.52449230 0.0337030065 0.5725336 0.6948152 0.6257220
#> [16,] 0.6928294 0.5335015 0.49899893 0.0640990319 0.4682144 0.4670711 0.4605605
#> [17,] 0.9391859 0.8598387 0.63117905 0.0049295527 0.4759836 0.4444418 0.3395994
#> [18,] 0.3695840 0.7000796 0.82159169 0.0589772016 0.4808539 0.5482986 0.6925557
#> [19,] 0.7058041 0.8772356 0.76985225 0.0420216199 0.7177629 0.6065477 0.4062192
#> [20,] 0.8308745 0.6310601 0.47866859 0.0190190664 0.4362339 0.4325018 0.4550594
#> [21,] 0.8381192 0.6544111 0.55599292 0.0140621123 0.4190105 0.3240454 0.1621245
#> [22,] 0.9870623 0.9004295 0.84083836 0.0048424443 0.7320291 0.6836095 0.6189992
#> [23,] 0.8010834 0.8908774 0.96393693 0.0211284780 0.8990695 0.8062917 0.7287981
#> [24,] 0.6426171 0.5728357 0.38833207 0.0845979692 0.4399109 0.4572347 0.3011169
#> [25,] 1.0225400 0.9720571 0.86585948 0.0004908812 0.7411270 0.6330206 0.4582121
#>               V33        V34        V35        V36        V37       V38
#>  [1,] 0.475632774 0.31843590 0.17553707 0.21300825 0.39588565 0.4576675
#>  [2,] 0.705412150 0.63395662 0.64846192 0.69363122 0.68261212 0.5653021
#>  [3,] 0.266917129 0.31555274 0.42581063 0.25357628 0.32253951 0.3602174
#>  [4,] 0.492658361 0.32958308 0.10267459 0.26422780 0.27763827 0.3682159
#>  [5,] 0.551884055 0.73868635 0.80032724 0.76208792 0.74125486 0.8377863
#>  [6,] 0.382789498 0.24172037 0.15352107 0.11771830 0.10506506 0.1480753
#>  [7,] 0.055086234 0.14327115 0.18539928 0.20560886 0.13926230 0.1643763
#>  [8,] 0.006937338 0.02530881 0.24611837 0.45196124 0.34189142 0.2262492
#>  [9,] 0.337030502 0.19751643 0.11287590 0.08968844 0.18726754 0.1703645
#> [10,] 0.278857293 0.28001635 0.35451158 0.13735072 0.10208782 0.2004127
#> [11,] 0.599493865 0.59242618 0.55051699 0.37730180 0.22879060 0.2944413
#> [12,] 0.716093830 0.51427863 0.42234691 0.29904485 0.12687697 0.2750298
#> [13,] 0.197643484 0.09434679 0.08095469 0.14256433 0.21095467 0.2381083
#> [14,] 0.496962187 0.50750675 0.50114678 0.46979279 0.42288844 0.3210945
#> [15,] 0.637991289 0.65441147 0.70743095 0.66715875 0.67665002 0.7491228
#> [16,] 0.548551848 0.52896341 0.55370699 0.54960599 0.44934063 0.2444264
#> [17,] 0.250285391 0.14013812 0.18627460 0.28689334 0.28269564 0.3052059
#> [18,] 0.603611111 0.88728650 0.81489786 0.75209350 0.78659268 0.4955570
#> [19,] 0.300218638 0.16730954 0.29082729 0.52022712 0.47548541 0.2541349
#> [20,] 0.507093887 0.60148549 0.67382274 0.64402993 0.46472303 0.3532505
#> [21,] 0.171173253 0.20823679 0.25686176 0.21058518 0.06475521 0.0770316
#> [22,] 0.561176052 0.58673540 0.58489623 0.49201772 0.44174652 0.3621184
#> [23,] 0.682659552 0.65268727 0.62844844 0.57612508 0.50287710 0.3668846
#> [24,] 0.239502109 0.16682296 0.46426525 0.67285631 0.66983912 0.4103146
#> [25,] 0.288696203 0.14552377 0.04184290 0.28576962 0.27267263 0.3599496
#>              V39          V4        V40        V41        V42        V43
#>  [1,] 0.46480178 0.064700340 0.52248202 0.38367556 0.39738382 0.28248833
#>  [2,] 0.46401542 0.054618399 0.51876033 0.48884544 0.41903370 0.45480258
#>  [3,] 0.26175043 0.057786524 0.11858282 0.57372185 0.65586174 0.54492714
#>  [4,] 0.55306900 0.044852119 0.43646706 0.25128615 0.36755007 0.42006707
#>  [5,] 0.77078302 0.127366409 0.57901217 0.39370507 0.37241853 0.37027360
#>  [6,] 0.16508097 0.065082404 0.14633111 0.22967351 0.25655582 0.16064011
#>  [7,] 0.32349450 0.070302335 0.33223770 0.21560210 0.12749713 0.27233124
#>  [8,] 0.23578846 0.027537172 0.19430044 0.24929954 0.37645524 0.35222093
#>  [9,] 0.16271114 0.052793696 0.17905790 0.19019764 0.22269421 0.19038950
#> [10,] 0.17803953 0.123337666 0.02639245 0.07851959 0.13559788 0.15866132
#> [11,] 0.32103614 0.037354650 0.40145860 0.36544868 0.36974723 0.30344544
#> [12,] 0.24650294 0.093599843 0.18935180 0.56061024 0.66421422 0.47751361
#> [13,] 0.21166889 0.059112848 0.24159166 0.25382626 0.27112055 0.25210075
#> [14,] 0.15183201 0.016700980 0.17394368 0.19022964 0.09224651 0.02497177
#> [15,] 0.77861855 0.010017300 0.73378626 0.51022949 0.44573149 0.30741448
#> [16,] 0.06553575 0.044897950 0.36012994 0.40557675 0.53511546 0.44646872
#> [17,] 0.26257995 0.008904171 0.15573943 0.13497886 0.15828237 0.14252854
#> [18,] 0.35091844 0.084004691 0.33929235 0.28240222 0.27409621 0.06108477
#> [19,] 0.26518190 0.048815243 0.21793467 0.29395732 0.27124209 0.18608159
#> [20,] 0.34176062 0.025689306 0.45659792 0.36045128 0.28974341 0.23603829
#> [21,] 0.13911151 0.015957265 0.20623719 0.27609245 0.26203777 0.19586969
#> [22,] 0.17616807 0.018629903 0.14433810 0.15969180 0.04280602 0.04672662
#> [23,] 0.25160835 0.032974154 0.18705729 0.15890261 0.13341041 0.15223273
#> [24,] 0.20934578 0.073496617 0.35718400 0.39739024 0.34089237 0.25021811
#> [25,] 0.39478386 0.021104573 0.22236387 0.05399914 0.22309500 0.16824901
#>               V44           V45          V46          V47         V48
#>  [1,] 0.150203985  5.843043e-02  0.040171473 -0.007057751 0.007155349
#>  [2,] 0.526555174  4.682369e-01  0.340634358  0.238024623 0.190056196
#>  [3,] 0.398642976  5.124495e-01  0.575617950  0.390226422 0.228629646
#>  [4,] 0.451801801  4.692439e-01  0.248598527  0.158360591 0.159490457
#>  [5,] 0.349117042  4.861872e-01  0.389738004  0.109073323 0.143268971
#>  [6,] 0.224985313  2.178054e-01  0.224605266  0.138511753 0.049735936
#>  [7,] 0.243705389  1.301644e-01  0.116963553  0.129469840 0.110052369
#>  [8,] 0.236949980  1.160029e-01  0.107639774  0.082053024 0.152460784
#>  [9,] 0.132143410  1.343742e-01  0.163543058  0.175517730 0.138309476
#> [10,] 0.125546265  6.266703e-02  0.010703614  0.063639847 0.066026904
#> [11,] 0.204631652  1.735438e-01  0.134492925  0.115489782 0.095844109
#> [12,] 0.468741052  5.778338e-01  0.570511650  0.381532990 0.181554520
#> [13,] 0.186015451  1.381731e-01  0.134002149  0.147027889 0.108730692
#> [14,] 0.014835724  3.980909e-02  0.031198915  0.016868671 0.021240989
#> [15,] 0.226988986  2.387853e-01  0.162286200  0.143240003 0.102536263
#> [16,] 0.262877603  1.637190e-01  0.112211364  0.052607241 0.065540352
#> [17,] 0.120335293  9.257196e-02  0.029969703  0.044622986 0.049686710
#> [18,] 0.170481564  1.556477e-01  0.167542387  0.123169218 0.074034979
#> [19,] 0.095342146  1.159011e-01  0.115121462  0.068468475 0.052628999
#> [20,] 0.195926069 -8.511669e-06 -0.003368410  0.019784236 0.015909162
#> [21,] 0.096677831  7.785309e-02  0.084637306  0.042406837 0.028268353
#> [22,] 0.007223895  1.499804e-03  0.007252646  0.017331763 0.031467787
#> [23,] 0.160255910  1.069037e-01  0.077674716  0.046460690 0.042143999
#> [24,] 0.210036398  2.145870e-01  0.168560150  0.164797452 0.151959747
#> [25,] 0.149149196  9.693912e-02  0.046454396  0.098867560 0.090391661
#>                V49           V5         V50           V51         V52
#>  [1,]  0.015267521  0.137005934 0.009940489  0.0097682076 0.016720118
#>  [2,]  0.104503131  0.079509659 0.031892269  0.0179260613 0.014590217
#>  [3,]  0.159534794  0.090245956 0.067451107  0.0673811844 0.032211165
#>  [4,]  0.088322843  0.084886235 0.025377923  0.0272561558 0.014711778
#>  [5,]  0.079942706  0.106910263 0.043934559  0.0194439050 0.026057330
#>  [6,]  0.049094450  0.087036024 0.010665572  0.0096340990 0.008449290
#>  [7,]  0.088184504  0.083935852 0.028556973  0.0326274085 0.007393836
#>  [8,]  0.100739891  0.136020787 0.018744133  0.0292105182 0.038091122
#>  [9,]  0.066488240  0.073166711 0.019964556  0.0187055253 0.013836440
#> [10,]  0.016393298  0.121028180 0.008943337  0.0270157302 0.012899633
#> [11,]  0.062695303  0.061775728 0.012919794  0.0044250238 0.010172027
#> [12,]  0.094401923  0.075750814 0.036213614  0.0455552588 0.050073623
#> [13,]  0.054386300  0.124582564 0.023550703  0.0189965346 0.011233520
#> [14,]  0.007924382  0.034622599 0.003580575  0.0012376390 0.003418413
#> [15,]  0.050015147  0.014048725 0.020977302  0.0141480659 0.010165355
#> [16,]  0.022467922  0.052217170 0.004734842  0.0006482161 0.006023081
#> [17,]  0.014503720  0.039339922 0.007236835  0.0002288863 0.008010697
#> [18,]  0.041360988  0.143278601 0.008361235  0.0175559671 0.005617613
#> [19,]  0.020472188  0.036538753 0.016402451  0.0090220638 0.013496961
#> [20,] -0.007829856  0.056818563 0.006806254  0.0197633409 0.005405220
#> [21,]  0.027303451  0.044853192 0.009978465  0.0111551839 0.008743519
#> [22,]  0.009764042  0.046949123 0.007858579  0.0056698631 0.003432287
#> [23,]  0.018980098  0.042583096 0.013543803  0.0078358258 0.007505824
#> [24,]  0.095756672  0.085342591 0.025375738  0.0163850549 0.015460011
#> [25,]  0.039847580 -0.001093982 0.012042607 -0.0015034107 0.002734026
#>               V53         V54           V55         V56         V57
#>  [1,] 0.016582628 0.020054700  0.0049824333 0.017017867 0.007288257
#>  [2,] 0.007477978 0.005849648  0.0056156751 0.007531891 0.005478007
#>  [3,] 0.015622714 0.023844225  0.0127000000 0.015914148 0.010430799
#>  [4,] 0.009924859 0.011980205  0.0137409585 0.007369156 0.007596266
#>  [5,] 0.027829263 0.023801950  0.0168647969 0.021245669 0.013863609
#>  [6,] 0.003768502 0.001478672 -0.0013728914 0.004170215 0.001825974
#>  [7,] 0.013234291 0.012720705  0.0277589488 0.009711251 0.012616697
#>  [8,] 0.020104376 0.037110461  0.0097647385 0.000231348 0.002760362
#>  [9,] 0.008487962 0.009665676  0.0048063877 0.005393949 0.004672679
#> [10,] 0.026237903 0.014154459  0.0176695646 0.018291458 0.017535803
#> [11,] 0.015194715 0.006202855  0.0056421357 0.012121334 0.006505061
#> [12,] 0.027255933 0.018563462  0.0117663408 0.006209781 0.019683471
#> [13,] 0.015440922 0.014674200  0.0052030229 0.011589639 0.006936942
#> [14,] 0.007457589 0.002023225  0.0048991969 0.005029990 0.001583934
#> [15,] 0.012180193 0.013768104  0.0116475037 0.006693351 0.005038589
#> [16,] 0.011387282 0.010019885  0.0100700063 0.013495160 0.007547682
#> [17,] 0.007400722 0.012393660  0.0113368901 0.006534444 0.003804963
#> [18,] 0.014437613 0.006891358  0.0194474074 0.007822551 0.014059424
#> [19,] 0.009243751 0.011603562  0.0139067168 0.011108403 0.011130681
#> [20,] 0.005893882 0.012786864  0.0036789799 0.007764678 0.008907711
#> [21,] 0.005611905 0.005776082  0.0066405984 0.003389317 0.003297903
#> [22,] 0.011124785 0.008326554  0.0044990208 0.005349281 0.003511672
#> [23,] 0.005969227 0.005322427  0.0036048253 0.003497012 0.005495494
#> [24,] 0.011506516 0.012579611  0.0074063317 0.006487295 0.013789942
#> [25,] 0.002209015 0.001336298  0.0007825835 0.004459546 0.003396342
#>                 V58         V59         V6          V60         V7         V8
#>  [1,]  0.0059283674 0.005781171 0.18268548  0.009617442 0.11937147 0.15555204
#>  [2,]  0.0048483199 0.003847650 0.10930673  0.003570388 0.13120192 0.16699039
#>  [3,]  0.0201974977 0.011171703 0.10554148  0.011757796 0.15244168 0.20493773
#>  [4,]  0.0070535293 0.006825219 0.10882671  0.004645768 0.11924954 0.15319748
#>  [5,]  0.0117794055 0.007724717 0.06965906  0.010270860 0.11762913 0.30962884
#>  [6,]  0.0074049752 0.002512202 0.09666199  0.002698579 0.12592546 0.17987879
#>  [7,]  0.0239406354 0.012039198 0.16083591  0.005160083 0.27538720 0.32584058
#>  [8,]  0.0208138222 0.013132005 0.25017033 -0.000434919 0.14990673 0.06800374
#>  [9,]  0.0059255573 0.005569430 0.10878395  0.005405848 0.12056557 0.12129971
#> [10,]  0.0144973777 0.011384528 0.10790600  0.012645087 0.11095215 0.10035981
#> [11,]  0.0078737986 0.010308879 0.08752587  0.004656043 0.11367115 0.06708858
#> [12,]  0.0143496753 0.013538542 0.07392024  0.003351124 0.02756044 0.10005601
#> [13,]  0.0169906847 0.006487838 0.15440186  0.005909965 0.12881522 0.17327804
#> [14,]  0.0041394325 0.009123811 0.06945867  0.005990193 0.08680296 0.06509780
#> [15,]  0.0059907645 0.006083411 0.04245363  0.003603837 0.06098737 0.08962648
#> [16,]  0.0019633087 0.004361333 0.08731657  0.014005530 0.09582160 0.14440722
#> [17,]  0.0064621736 0.005304097 0.04166960  0.003439974 0.05434870 0.04278073
#> [18,]  0.0058319342 0.018052675 0.21379358  0.007814650 0.20750173 0.18389630
#> [19,]  0.0065001274 0.006369078 0.10099375  0.004518582 0.15201708 0.15897470
#> [20,]  0.0022124898 0.007572628 0.08322317  0.003862363 0.10743623 0.04523874
#> [21,]  0.0020029458 0.002887068 0.05694293  0.002637641 0.05756419 0.05116471
#> [22,]  0.0034651263 0.002123932 0.08931514  0.002262336 0.08994477 0.08131313
#> [23,]  0.0047687298 0.006058369 0.06734226  0.003709932 0.05984021 0.08521761
#> [24,]  0.0099960060 0.008599312 0.11001502  0.007196766 0.18118535 0.22670993
#> [25,] -0.0002799461 0.002329428 0.01524864  0.002705490 0.08460385 0.08450841
#>               V9
#>  [1,] 0.14612664
#>  [2,] 0.28774171
#>  [3,] 0.24128032
#>  [4,] 0.26739960
#>  [5,] 0.56466423
#>  [6,] 0.25343873
#>  [7,] 0.32462466
#>  [8,] 0.11141095
#>  [9,] 0.16760837
#> [10,] 0.20208650
#> [11,] 0.10651693
#> [12,] 0.16489101
#> [13,] 0.24262829
#> [14,] 0.12407103
#> [15,] 0.03915076
#> [16,] 0.04683373
#> [17,] 0.02467703
#> [18,] 0.13589909
#> [19,] 0.18903547
#> [20,] 0.11541989
#> [21,] 0.10041648
#> [22,] 0.16934130
#> [23,] 0.10838414
#> [24,] 0.25003402
#> [25,] 0.07183322
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058