Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12       V13        V14
#>  [1,] 0.025523196 0.29631705 0.35056486 0.35613586 0.2702051 0.10231219
#>  [2,] 0.025610495 0.24438410 0.29661719 0.33427990 0.2556025 0.19287493
#>  [3,] 0.024877822 0.27947753 0.34872228 0.41505214 0.4228215 0.46174850
#>  [4,] 0.045619868 0.28897089 0.36395937 0.36513755 0.4821536 0.69901205
#>  [5,] 0.021451554 0.31371522 0.26618238 0.28417618 0.2369460 0.19278456
#>  [6,] 0.152776385 0.18544622 0.44118991 0.56522550 0.7352847 0.69611790
#>  [7,] 0.039572300 0.08211992 0.17360382 0.27685493 0.3947600 0.44145165
#>  [8,] 0.023782922 0.26657024 0.32794703 0.35992849 0.3604703 0.30455756
#>  [9,] 0.036644741 0.20538040 0.32152132 0.32277454 0.2915358 0.30421149
#> [10,] 0.014480302 0.18508064 0.17714165 0.13909020 0.1211183 0.21756517
#> [11,] 0.040880009 0.27505317 0.27934730 0.30080258 0.3422952 0.33396060
#> [12,] 0.030602595 0.23931018 0.39311621 0.48188832 0.5323395 0.52165555
#> [13,] 0.026771139 0.36150550 0.35884104 0.31936578 0.2969339 0.19916572
#> [14,] 0.012205293 0.10740041 0.10192090 0.09442098 0.1039250 0.12142409
#> [15,] 0.019591236 0.20179361 0.23679824 0.18953267 0.3075515 0.39820141
#> [16,] 0.011406688 0.28266623 0.30088339 0.32304309 0.3702856 0.49631786
#> [17,] 0.024987177 0.18665428 0.38663406 0.39261994 0.3986751 0.49398041
#> [18,] 0.012600000 0.42840000 0.30150000 0.12070000 0.3299000 0.57070000
#> [19,] 0.014638483 0.12842082 0.11734721 0.09744458 0.1276199 0.13515903
#> [20,] 0.040034659 0.21822200 0.24763398 0.34563996 0.3913595 0.48373184
#> [21,] 0.004881117 0.14509400 0.09763355 0.14674526 0.2209920 0.02950227
#> [22,] 0.019917740 0.14952926 0.15329891 0.11330231 0.1126937 0.12494985
#> [23,] 0.017151318 0.04517238 0.06854345 0.14097061 0.1558048 0.08696482
#> [24,] 0.020682275 0.61506315 0.64720927 0.75278571 0.5950932 0.54216808
#> [25,] 0.026898376 0.05032260 0.05510920 0.19265917 0.2637132 0.30131140
#>               V15       V16       V17       V18       V19          V2       V20
#>  [1,]  0.14014103 0.1907688 0.2259231 0.2419210 0.1737741 0.032128291 0.1853564
#>  [2,]  0.17099138 0.2088636 0.2000356 0.3185167 0.5355244 0.036558711 0.6565133
#>  [3,]  0.46506252 0.4590711 0.5686862 0.6364525 0.6973038 0.027572333 0.7228365
#>  [4,]  0.86470988 0.8780742 0.8010880 0.8829152 0.9605576 0.064108711 0.9081293
#>  [5,]  0.13719805 0.1717699 0.1789633 0.2251407 0.1836366 0.032315851 0.1279542
#>  [6,]  0.65319311 0.7654598 0.9359306 0.9046149 0.9456844 0.131893167 0.8078846
#>  [7,]  0.53572429 0.5936046 0.6993370 0.8149168 0.9407642 0.046703414 0.9502396
#>  [8,]  0.28655510 0.3083858 0.3437367 0.4168681 0.5645722 0.046569023 0.6883749
#>  [9,]  0.26201412 0.2311315 0.3194488 0.3189611 0.3779316 0.053713418 0.4200791
#> [10,]  0.34693378 0.5696613 0.7103330 0.7988589 0.8460172 0.016169087 0.8011819
#> [11,]  0.31883392 0.3198036 0.3066365 0.2055304 0.3049908 0.049330497 0.4561372
#> [12,]  0.61193449 0.7430698 0.8865056 0.9764082 0.9443515 0.049152690 0.8910249
#> [13,]  0.18848840 0.2046693 0.1947096 0.2555341 0.2582408 0.039655210 0.2134390
#> [14,]  0.17319695 0.1816004 0.1477211 0.1507597 0.1055042 0.015364067 0.1301298
#> [15,]  0.50498502 0.6006498 0.5592742 0.4686014 0.4431817 0.033315951 0.5137024
#> [16,]  0.68305473 0.8448772 0.9865302 0.9161923 0.8564142 0.019349400 0.8114844
#> [17,]  0.64839573 0.8089269 0.9794636 0.9971705 0.9427535 0.033906115 0.8385656
#> [18,]  0.69620000 0.9751000 1.0000000 0.9293000 0.6210000 0.051900000 0.4586000
#> [19,]  0.17335876 0.1546469 0.1467358 0.2224054 0.3019708 0.018738943 0.3316898
#> [20,]  0.67264500 0.7559035 0.9465661 0.9858097 0.9042933 0.043192599 0.8400565
#> [21,] -0.03649336 0.1526644 0.3344162 0.3369181 0.5205903 0.025131598 0.4125639
#> [22,]  0.15684729 0.3152626 0.4278610 0.4055035 0.3360918 0.007581488 0.4275393
#> [23,]  0.05684584 0.1564592 0.2394061 0.3680479 0.3272967 0.022940442 0.2175331
#> [24,]  0.66842427 0.7176790 0.7113577 0.7922088 0.9305627 0.056893720 0.8578140
#> [25,]  0.19076430 0.2213687 0.3123618 0.5004642 0.4810064 0.042577898 0.6317603
#>              V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.36991312 0.6219022 0.7825723 0.8371881 0.8906894 0.9461018 0.9852751
#>  [2,] 0.79381214 0.8295292 0.8084375 0.9075726 0.9074438 0.9576423 0.9753989
#>  [3,] 0.79389331 0.8680463 0.8553365 0.8753498 0.9155192 0.9324003 0.8867974
#>  [4,] 0.87900940 0.9104558 0.9247042 0.8314293 0.6963534 0.6529095 0.5292800
#>  [5,] 0.08306679 0.2074206 0.4141483 0.5681743 0.6851740 0.8152638 0.9144240
#>  [6,] 0.84276730 0.8029509 0.7343481 0.6637708 0.5065458 0.5506073 0.4616060
#>  [7,] 0.94250830 0.9986149 0.9865072 0.8539181 0.6801652 0.6146710 0.6554060
#>  [8,] 0.77702772 0.8397343 0.8967496 0.9735226 0.9757190 0.9963627 0.9802583
#>  [9,] 0.50254598 0.4467069 0.4133554 0.4491156 0.4478716 0.6594699 0.8564526
#> [10,] 0.79751724 0.8359917 0.8010897 0.4969510 0.2441399 0.3158767 0.5242138
#> [11,] 0.53168496 0.4812674 0.4810499 0.5005203 0.5352733 0.6109216 0.6642581
#> [12,] 0.79080047 0.6018683 0.4401969 0.3823136 0.3287332 0.1584069 0.1067615
#> [13,] 0.30063709 0.6047854 0.7798683 0.8176694 0.8338972 0.8375600 0.9045550
#> [14,] 0.21532490 0.2561773 0.3853264 0.6020101 0.8362484 0.9767731 0.8349008
#> [15,] 0.52498235 0.5274391 0.6455087 0.7964616 0.7756232 0.7619349 0.7555695
#> [16,] 0.83300307 0.8941482 0.7562085 0.6476787 0.7346221 0.8586120 0.6658629
#> [17,] 0.79955628 0.8445366 0.6132672 0.5387921 0.7234593 0.7863174 0.5745935
#> [18,] 0.50010000 0.5032000 0.7082000 0.8420000 0.8109000 0.7690000 0.8105000
#> [19,] 0.32386413 0.4562119 0.5557527 0.6077215 0.6005751 0.6330991 0.6320758
#> [20,] 0.76592600 0.8282223 0.6946742 0.6278030 0.6674459 0.7553661 0.6128936
#> [21,] 0.54047711 0.5303690 0.4777457 0.6224443 0.7502901 0.7025919 0.7941788
#> [22,] 0.55096294 0.6770098 0.7511780 0.7539195 0.8458272 0.9668634 0.9661005
#> [23,] 0.13701153 0.1407022 0.1943430 0.2304168 0.1228894 0.2094057 0.3816863
#> [24,] 0.77686528 0.7055523 0.4635872 0.4003308 0.5645866 0.2249140 0.5363756
#> [25,] 0.70825186 0.6912151 0.6922475 0.6230383 0.6728576 0.6865903 0.6595207
#>             V28        V29          V3       V30        V31       V32       V33
#>  [1,] 0.9978351 0.85685563 0.026602892 0.6692344 0.65803809 0.5634357 0.4291412
#>  [2,] 0.9488048 0.84005588 0.040176351 0.6265618 0.41970697 0.2945944 0.2401254
#>  [3,] 0.7460899 0.59932642 0.043190489 0.4283213 0.26967704 0.2345115 0.1603961
#>  [4,] 0.4033724 0.52557460 0.052242393 0.4059905 0.32912235 0.1614454 0.0668113
#>  [5,] 0.9362771 0.90241887 0.026930873 0.9322043 0.90962249 0.8214756 0.7109408
#>  [6,] 0.5960530 0.47751042 0.153174527 0.3101323 0.24154788 0.2252069 0.3109233
#>  [7,] 0.6945053 0.53842917 0.082508610 0.4303915 0.28978479 0.2017050 0.1996622
#>  [8,] 0.9236619 0.77865910 0.026997900 0.5875030 0.43746626 0.3788340 0.3009005
#>  [9,] 0.8329211 0.74649230 0.059171184 0.6946557 0.52380667 0.4298310 0.4011611
#> [10,] 0.7156553 0.80489142 0.003686493 0.7277137 0.48349923 0.4343917 0.4068075
#> [11,] 0.7809784 0.80528884 0.066522274 0.7759222 0.74895631 0.7010355 0.5585356
#> [12,] 0.1103727 0.04795294 0.068959802 0.1987846 0.26316878 0.3550812 0.3510603
#> [13,] 0.9867425 0.80635930 0.049278018 0.5581159 0.63910774 0.6304711 0.4944381
#> [14,] 0.6082478 0.49621637 0.012640088 0.3823062 0.29416417 0.1508278 0.1728485
#> [15,] 0.7662697 0.66244325 0.027716565 0.6131933 0.50853177 0.3198636 0.2792317
#> [16,] 0.5093892 0.61030705 0.018647338 0.6815002 0.58756657 0.4211926 0.4032630
#> [17,] 0.3397047 0.50717224 0.039460643 0.6238151 0.55516427 0.3934673 0.3775785
#> [18,] 0.6203000 0.23560000 0.062100000 0.2595000 0.62990000 0.6762000 0.2903000
#> [19,] 0.7421248 0.87009455 0.023411835 0.8739199 0.76370667 0.6724955 0.7202307
#> [20,] 0.6303783 0.89279135 0.051875102 0.8616882 0.55409366 0.2699700 0.3426063
#> [21,] 0.8413198 0.69594822 0.045727892 0.3331692 0.08758636 0.2313094 0.4663838
#> [22,] 0.8989458 0.77849221 0.012781883 0.7309605 0.70831842 0.6259492 0.5763325
#> [23,] 0.6250102 0.78702010 0.037745279 0.7123301 0.60720669 0.4042459 0.5477074
#> [24,] 0.8876395 0.59459190 0.117706327 0.8534691 0.89494469 0.5181271 0.1455068
#> [25,] 0.5759249 0.43069089 0.059487352 0.3387276 0.45262754 0.5735158 0.5071610
#>              V34       V35        V36        V37        V38        V39
#>  [1,] 0.18101644 0.1003104 0.01673653 0.07604031 0.10275224 0.13824778
#>  [2,] 0.21786917 0.1617265 0.04226516 0.06500662 0.10912241 0.09221335
#>  [3,] 0.05440198 0.0729192 0.17451753 0.24631222 0.32781658 0.40567314
#>  [4,] 0.14903397 0.1906926 0.18527381 0.17083632 0.10731126 0.26067169
#>  [5,] 0.64371227 0.6562011 0.70551823 0.69545616 0.57041917 0.47537614
#>  [6,] 0.32236130 0.3230388 0.06639129 0.14885240 0.29993579 0.34457595
#>  [7,] 0.25847755 0.3440607 0.33180636 0.29797081 0.35873234 0.37430289
#>  [8,] 0.24153612 0.1970400 0.20782421 0.23335606 0.22308669 0.23018976
#>  [9,] 0.34737950 0.2172289 0.16773777 0.25637470 0.22147670 0.16490945
#> [10,] 0.34111503 0.2770565 0.21282079 0.21013455 0.33212944 0.32844223
#> [11,] 0.43396923 0.3401773 0.29610906 0.24622202 0.35659312 0.42253899
#> [12,] 0.41720891 0.4438346 0.23849842 0.20861122 0.36603383 0.27543506
#> [13,] 0.24930615 0.1181814 0.06301417 0.17786500 0.19407408 0.20882922
#> [14,] 0.20695100 0.2478484 0.18827099 0.06784181 0.06328454 0.15177887
#> [15,] 0.20703678 0.3612231 0.51931934 0.41033856 0.21533716 0.18138775
#> [16,] 0.39090155 0.1779791 0.18493089 0.24025866 0.13945591 0.14948543
#> [17,] 0.35229532 0.1715713 0.13896057 0.20218281 0.17403965 0.13917989
#> [18,] 0.43930000 0.8529000 0.71800000 0.48010000 0.58560000 0.49930000
#> [19,] 0.73804312 0.6780061 0.59531389 0.42616690 0.29811719 0.22979178
#> [20,] 0.46331275 0.2998944 0.10791039 0.20096760 0.17958701 0.06501440
#> [21,] 0.73490676 0.8286503 0.83862886 0.62307285 0.55480630 0.45370332
#> [22,] 0.56844846 0.5908914 0.56936701 0.53557236 0.45038476 0.25633092
#> [23,] 0.54804232 0.3967207 0.51974668 0.62593169 0.68300431 0.70193588
#> [24,] 0.21515843 0.3842102 0.26113007 0.60177678 0.66105327 0.51423777
#> [25,] 0.42438526 0.4739333 0.60670573 0.60745064 0.48605464 0.35760861
#>                V4        V40        V41        V42        V43        V44
#>  [1,] 0.041031787 0.08292678 0.13681998 0.10476565 0.16490703 0.13014674
#>  [2,] 0.044201957 0.10385407 0.05799990 0.12066795 0.20869576 0.22006948
#>  [3,] 0.049988353 0.37262196 0.28592444 0.23749118 0.21366254 0.19951697
#>  [4,] 0.036146418 0.29724396 0.17332313 0.14439099 0.25259606 0.22345921
#>  [5,] 0.048299573 0.51055572 0.48131837 0.43382809 0.45723263 0.52384751
#>  [6,] 0.162583154 0.25362808 0.09073209 0.24539590 0.37379910 0.17602122
#>  [7,] 0.095287165 0.32089803 0.35437279 0.44101879 0.35669751 0.24615268
#>  [8,] 0.020256147 0.22855154 0.12257989 0.12233673 0.21452194 0.25327594
#>  [9,] 0.100024513 0.29771661 0.36784955 0.38292271 0.36537651 0.22787735
#> [10,] 0.028586933 0.24809958 0.13278751 0.17944516 0.29162717 0.26471908
#> [11,] 0.060944621 0.36994935 0.51676287 0.56694758 0.48340763 0.48132055
#> [12,] 0.072644073 0.13884028 0.18198405 0.21609153 0.16130740 0.08062642
#> [13,] 0.075095202 0.15574431 0.19126800 0.08771479 0.07575267 0.12067968
#> [14,] 0.013041342 0.21773094 0.28296313 0.27529217 0.20555247 0.12368388
#> [15,] 0.013003018 0.22315232 0.25691359 0.23172140 0.20044719 0.14550851
#> [16,] 0.048404117 0.29395920 0.18390705 0.18764942 0.12336450 0.11623962
#> [17,] 0.025605853 0.25722471 0.23327515 0.21783603 0.09026876 0.13993671
#> [18,] 0.051800000 0.28660000 0.06010000 0.11670000 0.27370000 0.28120000
#> [19,] 0.029534598 0.27174970 0.27286691 0.25905145 0.25368501 0.22776110
#> [20,] 0.013899142 0.14676004 0.15477839 0.19129184 0.15596509 0.09562904
#> [21,] 0.009913028 0.23935090 0.02399833 0.29740594 0.30229324 0.44929709
#> [22,] 0.029425577 0.17103787 0.18348499 0.13575226 0.14525490 0.11560965
#> [23,] 0.016438415 0.59344108 0.44340990 0.35882561 0.24953695 0.12619761
#> [24,] 0.102270213 0.45815794 0.46567187 0.23631559 0.18835239 0.20652464
#> [25,] 0.067519175 0.45860852 0.45689161 0.40391280 0.35369866 0.25973183
#>              V45         V46        V47        V48         V49         V5
#>  [1,] 0.08186179  0.09122053 0.14308687 0.14094937 0.083212040 0.05857447
#>  [2,] 0.21559861  0.23920232 0.18438225 0.10126440 0.054031586 0.07422676
#>  [3,] 0.18858555  0.17427391 0.14797360 0.09131336 0.058637426 0.12101164
#>  [4,] 0.12239682  0.12608814 0.11969343 0.08012541 0.058723967 0.02887926
#>  [5,] 0.49221681  0.35702395 0.25895493 0.20223980 0.110829066 0.07208483
#>  [6,] 0.05892784  0.10643221 0.06713746 0.05562526 0.056512089 0.17745813
#>  [7,] 0.18129877  0.12264606 0.07310341 0.05914794 0.053006044 0.18726667
#>  [8,] 0.22107318  0.19040488 0.11959798 0.05975364 0.030869593 0.04553172
#>  [9,] 0.21978121  0.17863782 0.14725229 0.14287407 0.090378140 0.10181576
#> [10,] 0.13036137  0.13393554 0.11971935 0.07596266 0.055418302 0.03862473
#> [11,] 0.57424034  0.48269425 0.33313610 0.22462051 0.119748542 0.06818010
#> [12,] 0.06467210  0.03921335 0.04768476 0.03521875 0.016488807 0.07714599
#> [13,] 0.05926225  0.06318521 0.07189091 0.10442376 0.069360382 0.10103344
#> [14,] 0.09890756  0.10170227 0.05198336 0.03527389 0.029386010 0.03077365
#> [15,] 0.12407021  0.11961082 0.06884302 0.05925660 0.033618347 0.04141979
#> [16,] 0.18854821  0.20124342 0.17899384 0.13938646 0.071511327 0.08205815
#> [17,] 0.15709855  0.17781360 0.18647515 0.13135395 0.048306301 0.11961377
#> [18,] 0.20780000  0.06600000 0.04910000 0.03450000 0.017200000 0.10720000
#> [19,] 0.17452903  0.12745788 0.09259080 0.06197512 0.030344692 0.04037937
#> [20,] 0.21123398  0.24614746 0.17684052 0.10395153 0.063697446 0.01814911
#> [21,] 0.23442727  0.10930395 0.09607661 0.14782652 0.039138987 0.09892717
#> [22,] 0.08942497  0.08800891 0.05011443 0.03282578 0.021723810 0.04666106
#> [23,] 0.06372833  0.07709559 0.13394501 0.10769544 0.053869897 0.03555802
#> [24,] 0.18411749 -0.02067922 0.14086114 0.06684448 0.006238696 0.09367972
#> [25,] 0.14611669  0.12002503 0.06384371 0.03110033 0.025271819 0.08843677
#>               V50         V51          V52         V53         V54         V55
#>  [1,] 0.021810008 0.017058618 1.045686e-02 0.008685691 0.017336346 0.004862497
#>  [2,] 0.017338716 0.020970049 1.203267e-02 0.005274832 0.005510274 0.003778347
#>  [3,] 0.028552006 0.017984660 1.621120e-02 0.014475828 0.019381929 0.010781409
#>  [4,] 0.019244654 0.033533661 1.183495e-02 0.009829637 0.019006261 0.012513899
#>  [5,] 0.031318756 0.017837814 1.542361e-02 0.009272845 0.006583731 0.005566108
#>  [6,] 0.030486307 0.019483638 1.424520e-02 0.016195744 0.038517894 0.007597116
#>  [7,] 0.011892702 0.014519075 1.898130e-02 0.011896826 0.025381184 0.009190644
#>  [8,] 0.007703779 0.016887713 9.435489e-03 0.011090671 0.009352690 0.005977576
#>  [9,] 0.018428872 0.018420867 1.570946e-02 0.010365631 0.009900385 0.008805293
#> [10,] 0.018919919 0.012537575 1.472214e-02 0.006285263 0.004185675 0.005446621
#> [11,] 0.040635649 0.037394532 2.876123e-02 0.016291652 0.017558618 0.014638705
#> [12,] 0.011085578 0.018045918 1.482730e-02 0.017783890 0.019244113 0.021386595
#> [13,] 0.023990229 0.014246136 1.034745e-02 0.005594484 0.004774795 0.004734739
#> [14,] 0.010792168 0.005417406 8.678270e-03 0.003508590 0.004085213 0.006213044
#> [15,] 0.018019455 0.014405385 1.390854e-02 0.009159712 0.010219972 0.009653286
#> [16,] 0.018577874 0.012533962 1.260532e-02 0.008089364 0.007730360 0.006923157
#> [17,] 0.020796248 0.009487771 3.428231e-03 0.008910480 0.007683918 0.007993606
#> [18,] 0.028700000 0.002700000 2.080000e-02 0.004800000 0.019900000 0.012600000
#> [19,] 0.017467634 0.015728329 1.061863e-02 0.009533922 0.008559317 0.005123813
#> [20,] 0.039974307 0.011697097 3.140613e-03 0.003453884 0.011146451 0.012938040
#> [21,] 0.034369050 0.024227331 8.030681e-05 0.005729502 0.014615837 0.003707076
#> [22,] 0.013198095 0.004671536 6.824012e-03 0.003767677 0.003719013 0.006348160
#> [23,] 0.019914143 0.010640690 1.396617e-02 0.020874775 0.011706720 0.014158760
#> [24,] 0.006806564 0.001627962 2.251028e-02 0.014259147 0.007420403 0.018052251
#> [25,] 0.021747267 0.009337252 1.279070e-02 0.013149355 0.011688195 0.008675234
#>               V56          V57         V58         V59         V6         V60
#>  [1,] 0.009036586 0.0051572000 0.008801300 0.014601966 0.16132101 0.011670830
#>  [2,] 0.003613975 0.0030180366 0.002178178 0.002874091 0.06955106 0.006159084
#>  [3,] 0.009483369 0.0100716775 0.011422811 0.010773327 0.16680618 0.008646260
#>  [4,] 0.008779599 0.0103678710 0.014409637 0.009795984 0.14325111 0.011370208
#>  [5,] 0.007767860 0.0060471008 0.004992430 0.003546746 0.09947414 0.003022485
#>  [6,] 0.007793236 0.0003663303 0.011415004 0.016094759 0.13488904 0.005897032
#>  [7,] 0.009816411 0.0113202255 0.018895125 0.013578578 0.20415720 0.007363855
#>  [8,] 0.005504324 0.0040779057 0.004681120 0.007238968 0.09286525 0.004340406
#>  [9,] 0.009640053 0.0058102595 0.011865316 0.011340291 0.08749894 0.006238233
#> [10,] 0.005293690 0.0026878108 0.005744788 0.003410035 0.07979700 0.003946655
#> [11,] 0.009184098 0.0104971127 0.009854376 0.012113753 0.10372342 0.006914489
#> [12,] 0.012084844 0.0080565463 0.013497529 0.009198225 0.04678841 0.007129401
#> [13,] 0.009561003 0.0053395458 0.005158093 0.006481796 0.10683897 0.009428823
#> [14,] 0.003355902 0.0031335392 0.003064709 0.001296877 0.04521593 0.002893225
#> [15,] 0.008537830 0.0084528384 0.005837535 0.006630746 0.10765782 0.003032259
#> [16,] 0.006182161 0.0076457980 0.001970840 0.005127959 0.09819794 0.001950600
#> [17,] 0.008829458 0.0037506588 0.003604846 0.001985483 0.12492577 0.006572909
#> [18,] 0.002200000 0.0037000000 0.003400000 0.011400000 0.25870000 0.007700000
#> [19,] 0.006018469 0.0069780692 0.004940824 0.008291132 0.06371717 0.005613329
#> [20,] 0.008589650 0.0058525931 0.005680743 0.003977996 0.10626115 0.009030750
#> [21,] 0.019011494 0.0196437291 0.007749357 0.009136977 0.09622002 0.001299438
#> [22,] 0.004520106 0.0073244592 0.007399681 0.005068553 0.08011386 0.005550730
#> [23,] 0.005185899 0.0043907831 0.010325585 0.013654473 0.07180852 0.008998567
#> [24,] 0.024347749 0.0334287916 0.017350972 0.010231517 0.24946479 0.007559645
#> [25,] 0.010484946 0.0110079161 0.007054763 0.006811742 0.13632478 0.010179343
#>               V7         V8         V9
#>  [1,] 0.17476167 0.17056784 0.22831808
#>  [2,] 0.09830985 0.13385152 0.18861560
#>  [3,] 0.17794407 0.19840695 0.26575871
#>  [4,] 0.21074390 0.25031101 0.28679328
#>  [5,] 0.12664416 0.14384606 0.28055207
#>  [6,] 0.17673249 0.19726714 0.14116083
#>  [7,] 0.15102751 0.10572148 0.10288045
#>  [8,] 0.11599901 0.12908986 0.21351686
#>  [9,] 0.09774610 0.13316951 0.20103833
#> [10,] 0.09572065 0.06279105 0.14896914
#> [11,] 0.09487745 0.13607485 0.24474542
#> [12,] 0.08650722 0.07984945 0.10678256
#> [13,] 0.12818353 0.20752807 0.28151466
#> [14,] 0.06303972 0.06449698 0.10678573
#> [15,] 0.14524503 0.18656059 0.18607073
#> [16,] 0.16020035 0.19726330 0.25632248
#> [17,] 0.12943063 0.09960998 0.12959783
#> [18,] 0.23040000 0.20670000 0.34160000
#> [19,] 0.08097880 0.07555194 0.08765985
#> [20,] 0.11775385 0.10422854 0.20981338
#> [21,] 0.12337116 0.10892181 0.26951570
#> [22,] 0.09951464 0.09811061 0.11844416
#> [23,] 0.05188847 0.06411889 0.05747115
#> [24,] 0.26044739 0.38852211 0.55911028
#> [25,] 0.12269096 0.09361617 0.04608308
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623