Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1          V10        V11        V12        V13        V14
#>  [1,] 0.005220048  0.108718180 0.15652073 0.23665493 0.21547807 0.35768576
#>  [2,] 0.034587202  0.212309893 0.21303647 0.25885763 0.36061499 0.34007728
#>  [3,] 0.089707004  0.543068534 0.41038810 0.17896761 0.15579086 0.29988191
#>  [4,] 0.041025802 -0.070719288 0.08985019 0.25466000 0.33120546 0.40297202
#>  [5,] 0.033774618  0.197150372 0.29972767 0.31624757 0.29266331 0.30100004
#>  [6,] 0.026111183  0.252659755 0.29744957 0.34696189 0.28390145 0.22465938
#>  [7,] 0.044358856  0.253718231 0.35642515 0.41938062 0.40231917 0.46027078
#>  [8,] 0.032115770  0.236050784 0.34671453 0.45257294 0.48117279 0.42843085
#>  [9,] 0.089174595  0.385807230 0.44361484 0.45242688 0.57632606 0.75821499
#> [10,] 0.051993705  0.018457607 0.15245238 0.19209061 0.23349114 0.29337129
#> [11,] 0.020507776  0.195115968 0.24659011 0.29830098 0.33914400 0.40317365
#> [12,] 0.045050972  0.298938854 0.32143657 0.28754202 0.22552472 0.11198923
#> [13,] 0.043378958  0.222572801 0.27452262 0.30701485 0.32552460 0.28553987
#> [14,] 0.025828446  0.478244200 0.55632991 0.52182395 0.46483458 0.34724468
#> [15,] 0.012876346  0.080052710 0.10679179 0.17872716 0.18973538 0.23927958
#> [16,] 0.019760146  0.117828451 0.11825530 0.07894918 0.18003090 0.30932832
#> [17,] 0.033939492  0.158243156 0.15974606 0.13486491 0.23353813 0.33077359
#> [18,] 0.020618711  0.162315094 0.16983316 0.21135736 0.30749806 0.13253864
#> [19,] 0.029300000  0.086000000 0.04140000 0.04720000 0.08350000 0.09380000
#> [20,] 0.007833497  0.157513397 0.16009104 0.16747415 0.20322337 0.29222208
#> [21,] 0.012486417  0.059063654 0.06416193 0.04931710 0.09757995 0.11253074
#> [22,] 0.011980871  0.095901737 0.10028575 0.08365059 0.11011364 0.13572718
#> [23,] 0.017376549  0.119437542 0.11948171 0.06406055 0.06962355 0.10146136
#> [24,] 0.004785261  0.070074176 0.02665873 0.15143039 0.12817301 0.07961645
#> [25,] 0.022165571  0.006631579 0.11863024 0.37395701 0.34301463 0.34239090
#>              V15        V16        V17       V18        V19          V2
#>  [1,] 0.50054027 0.56454017 0.70270261 0.8281287 0.93004689 0.010535838
#>  [2,] 0.31934699 0.43490376 0.46908388 0.5080566 0.62003543 0.048214830
#>  [3,] 0.43457528 0.45446982 0.18083394 0.1357118 0.29401851 0.104528814
#>  [4,] 0.47617751 0.57453352 0.70072655 0.7558633 0.82045878 0.053766654
#>  [5,] 0.25988407 0.22343680 0.27118005 0.2951703 0.35451282 0.048587497
#>  [6,] 0.16326621 0.19836923 0.18944525 0.2993430 0.53465964 0.032329801
#>  [7,] 0.49516834 0.40745101 0.42253034 0.5424733 0.71473890 0.039875194
#>  [8,] 0.50995263 0.65202933 0.81449125 0.9498843 1.02206695 0.062188273
#>  [9,] 0.98540456 1.02661186 0.80729937 0.8685734 0.95732920 0.107053242
#> [10,] 0.39304851 0.48539858 0.64617099 0.8126643 0.94025001 0.031753168
#> [11,] 0.46192483 0.53465698 0.62178485 0.7079644 0.77704960 0.019106693
#> [12,] 0.09964969 0.16487214 0.17686651 0.2374637 0.20700969 0.039808976
#> [13,] 0.24889391 0.28873332 0.30836515 0.2407353 0.30312771 0.062247687
#> [14,] 0.25018099 0.26563016 0.28779977 0.3316232 0.50748693 0.031958965
#> [15,] 0.22037064 0.38387650 0.49373943 0.4176297 0.35752604 0.020300511
#> [16,] 0.53497648 0.73444356 0.79584820 0.7855462 0.85430264 0.016031884
#> [17,] 0.41170760 0.43069275 0.42899130 0.5386056 0.63313461 0.041893014
#> [18,] 0.21369372 0.27753851 0.31412691 0.3592437 0.35055889 0.051631153
#> [19,] 0.14660000 0.08090000 0.11790000 0.2179000 0.33260000 0.064400000
#> [20,] 0.27637067 0.30705496 0.47120697 0.6759083 0.76109268 0.011071395
#> [21,] 0.15751647 0.23167797 0.29130790 0.3546608 0.43277750 0.012771244
#> [22,] 0.19195061 0.20632372 0.22367493 0.2448977 0.28667794 0.011834965
#> [23,] 0.10949926 0.13303467 0.18496071 0.2489571 0.27476074 0.011256391
#> [24,] 0.11290678 0.10994470 0.08269223 0.1632086 0.07022246 0.009807653
#> [25,] 0.04346302 0.05228268 0.21062603 0.4951731 0.40177224 0.009390454
#>             V20       V21       V22        V23        V24         V25
#>  [1,] 0.9715430 0.8999217 0.8731353 0.94507464 0.83258597  0.48488170
#>  [2,] 0.7396336 0.7075012 0.7882356 0.92608834 0.92825589  0.94064169
#>  [3,] 0.8809801 0.9603858 0.4378548 0.12749889 0.17869296  0.21860994
#>  [4,] 0.9291839 0.8585900 0.7891476 0.83353377 0.98288352  0.91907534
#>  [5,] 0.3752442 0.4109916 0.3966925 0.48409983 0.60168882  0.64752838
#>  [6,] 0.6534073 0.7789267 0.8031149 0.78496722 0.89277158  0.90524471
#>  [7,] 0.7203416 0.7311770 0.8365298 0.83887371 0.80750198  0.69483253
#>  [8,] 0.9930991 0.9083319 0.6821426 0.52712074 0.47048304  0.31121374
#>  [9,] 1.0614034 0.9861671 0.9071068 0.99141354 0.95663927  0.71597667
#> [10,] 0.9847487 0.9760659 0.9827513 0.99493494 0.92983225  0.81942179
#> [11,] 0.8551198 0.9321268 0.9533672 0.92449298 0.93052191  0.93167926
#> [12,] 0.1706417 0.3159314 0.5943332 0.76191665 0.82991573  0.92546225
#> [13,] 0.3901574 0.4695928 0.5113466 0.55708508 0.58495308  0.57871489
#> [14,] 0.6708462 0.7948195 0.8488301 0.81118455 0.83043729  0.79632471
#> [15,] 0.3538594 0.4775282 0.5103115 0.50627230 0.58890082  0.71790308
#> [16,] 0.8657582 0.9537642 0.9681916 0.93979383 0.93087016  0.88849249
#> [17,] 0.8199056 0.8090098 0.6552840 0.60708994 0.68551985  0.65463064
#> [18,] 0.1693811 0.2268191 0.1821233 0.19255853 0.21190528  0.26875622
#> [19,] 0.3258000 0.2111000 0.2302000 0.33610000 0.42590000  0.46090000
#> [20,] 0.8988569 0.9783073 0.9452449 0.86484632 0.77010324  0.69055651
#> [21,] 0.4357177 0.4409986 0.4080304 0.34622323 0.42410083  0.44468422
#> [22,] 0.3231551 0.3555881 0.3999417 0.54852911 0.71820901  0.82879896
#> [23,] 0.3788672 0.4923269 0.6429809 0.78737453 0.80218256  0.77064900
#> [24,] 0.1610283 0.2805466 0.1252518 0.06211742 0.09809849 -0.00599799
#> [25,] 0.6222323 0.7059300 0.7296222 0.79500296 0.85562923  1.02483035
#>              V26        V27        V28       V29          V3       V30
#>  [1,] 0.36608859 0.33337570 0.44244288 0.4978765 0.032389847 0.4453055
#>  [2,] 1.01173297 0.96733862 0.92554598 0.7395511 0.025911708 0.6109546
#>  [3,] 0.45017283 0.31239848 0.50791369 0.4821103 0.088683357 0.7301526
#>  [4,] 0.63611029 0.38535846 0.27851923 0.2920721 0.069514387 0.2113413
#>  [5,] 0.68906010 0.82077445 0.87885236 0.8332551 0.060687948 0.6638125
#>  [6,] 0.98113089 1.01085377 0.97456852 0.8706264 0.042744198 0.6410627
#>  [7,] 0.85658499 0.92054370 0.68318374 0.5665002 0.074030174 0.5116912
#>  [8,] 0.07818728 0.03581396 0.02751071 0.0401575 0.073553244 0.1837316
#>  [9,] 0.57122495 0.58820755 0.60282299 0.6155952 0.055565568 0.4611461
#> [10,] 0.68453817 0.54639332 0.47837047 0.4541095 0.042946726 0.3475170
#> [11,] 0.92169392 0.89096447 0.83691850 0.7202436 0.023319160 0.5616474
#> [12,] 0.98553277 0.97021411 0.98859343 0.8449869 0.028571389 0.6712993
#> [13,] 0.62328879 0.68850200 0.79173515 0.7933656 0.060244154 0.8106182
#> [14,] 0.81762864 0.85438928 0.82872455 0.6650489 0.045458521 0.4668953
#> [15,] 0.73723764 0.73236856 0.77475350 0.8094146 0.013873328 0.8569449
#> [16,] 0.91528403 0.87244342 0.67787687 0.4761792 0.032971550 0.4486105
#> [17,] 0.51157595 0.45080046 0.56107434 0.4707344 0.041835950 0.5570984
#> [18,] 0.29790502 0.15495888 0.49096072 0.7332014 0.039898656 0.7235464
#> [19,] 0.26060000 0.08740000 0.28620000 0.5606000 0.039000000 0.8344000
#> [20,] 0.55175141 0.32081121 0.08369467 0.3548230 0.007447444 0.2725364
#> [21,] 0.49482201 0.69282128 0.90877564 1.0001017 0.013012987 0.9103422
#> [22,] 0.85567635 0.82293297 0.73267227 0.5518222 0.013648577 0.4187299
#> [23,] 0.81892679 0.82242145 0.83517802 0.8911189 0.011316492 0.8054076
#> [24,] 0.14559170 0.34367983 0.51155994 0.5556294 0.015791167 0.5833545
#> [25,] 0.88349893 0.63687679 0.49136447 0.5013720 0.075033525 0.4198645
#>             V31       V32       V33        V34       V35        V36        V37
#>  [1,] 0.3037930 0.3700710 0.3948089 0.26746443 0.2345648 0.23867228 0.29693393
#>  [2,] 0.5670132 0.5555684 0.3994259 0.26516976 0.1695583 0.15496012 0.25161179
#>  [3,] 0.3923411 0.4255864 0.5241771 0.66810887 0.8121230 0.83523335 0.80935279
#>  [4,] 0.2755258 0.3993117 0.4748737 0.49653299 0.5077784 0.41659583 0.36180527
#>  [5,] 0.5691046 0.4929128 0.4513468 0.37530237 0.2300116 0.17767856 0.21837943
#>  [6,] 0.4216587 0.2835627 0.2861511 0.24278863 0.1519862 0.01412403 0.04009893
#>  [7,] 0.4141553 0.2909220 0.1383901 0.10526997 0.1328359 0.09256161 0.06269509
#>  [8,] 0.2326557 0.3256442 0.3548941 0.49498216 0.5224328 0.27386776 0.31314955
#>  [9,] 0.3431873 0.2172846 0.0258888 0.12042754 0.1673880 0.13888212 0.14428473
#> [10,] 0.2687166 0.2424941 0.2714800 0.38905440 0.4031136 0.42484766 0.44219467
#> [11,] 0.3918452 0.2856830 0.2024502 0.09763423 0.1056710 0.14722183 0.20044557
#> [12,] 0.7493048 0.6999874 0.5015785 0.19458568 0.1085942 0.10782827 0.12174995
#> [13,] 0.8214833 0.7513833 0.5877970 0.55369456 0.5401210 0.42171919 0.32994929
#> [14,] 0.3842934 0.3863260 0.4141565 0.38987946 0.2332255 0.31577771 0.31954306
#> [15,] 0.7276507 0.6133931 0.6721781 0.71240023 0.7738677 0.67226604 0.55260619
#> [16,] 0.4593708 0.3802890 0.3411302 0.31210709 0.2006811 0.09753463 0.20915677
#> [17,] 0.5330074 0.4152491 0.4346860 0.43367668 0.4003377 0.50941072 0.47948821
#> [18,] 0.6072751 0.6212946 0.6582100 0.62269287 0.6206867 0.76157758 0.84577909
#> [19,] 0.8096000 0.7250000 0.8048000 0.94350000 1.0000000 0.89600000 0.55160000
#> [20,] 0.2386795 0.1380052 0.1810789 0.25121535 0.2803426 0.40678473 0.39719384
#> [21,] 0.8309062 0.7586737 0.7277876 0.61289867 0.5423122 0.53246557 0.49716512
#> [22,] 0.3711514 0.2992455 0.2414905 0.25799766 0.2632069 0.26929940 0.23907667
#> [23,] 0.7125487 0.6199008 0.6008009 0.64218587 0.6086393 0.51916993 0.38452176
#> [24,] 0.6489242 0.6010767 0.6727230 0.91363800 1.0275828 0.97539395 0.84866747
#> [25,] 0.4889550 0.5436662 0.6330541 0.56113272 0.5276727 0.51198344 0.41569547
#>              V38        V39           V4        V40        V41        V42
#>  [1,] 0.46985189 0.34140213  0.103056351 0.20433900 0.13743929 0.19174022
#>  [2,] 0.37449438 0.39642086  0.046003873 0.30982065 0.28822575 0.25686464
#>  [3,] 0.88722352 0.74586631  0.098379259 0.50309712 0.36443181 0.32728894
#>  [4,] 0.42021299 0.42004015  0.052594391 0.42558957 0.34678671 0.34189168
#>  [5,] 0.15405638 0.20090016  0.091695607 0.29842304 0.33156414 0.37218715
#>  [6,] 0.10513888 0.09963561  0.036892837 0.11510202 0.04591470 0.12212130
#>  [7,] 0.15090736 0.22539505  0.099828233 0.19505385 0.24612451 0.27596863
#>  [8,] 0.43092952 0.24393541  0.084108428 0.25969398 0.41003323 0.32123393
#>  [9,] 0.07379545 0.15212946 -0.014229150 0.26707823 0.18452508 0.11992953
#> [10,] 0.38045945 0.43806252  0.056091176 0.35048125 0.33226848 0.40426979
#> [11,] 0.25615985 0.25575644  0.056673334 0.24865944 0.22631078 0.23068348
#> [12,] 0.15732840 0.19105134  0.049980054 0.05384639 0.14945531 0.04689070
#> [13,] 0.38707054 0.40788702  0.060184339 0.39577570 0.57110841 0.60325231
#> [14,] 0.34023645 0.49655423  0.048982842 0.42862079 0.19212190 0.27769161
#> [15,] 0.45368994 0.27209531  0.010583810 0.15742723 0.16814534 0.15314644
#> [16,] 0.15891394 0.10374214  0.040235141 0.13680037 0.10191891 0.05444272
#> [17,] 0.47700633 0.51158684  0.057656992 0.48274168 0.35209231 0.21685849
#> [18,] 0.93133287 0.97799876  0.020723615 0.91349135 0.85610073 0.76787652
#> [19,] 0.30370000 0.23380000  0.017300000 0.23820000 0.33180000 0.38210000
#> [20,] 0.19771465 0.13529220  0.025357106 0.13219294 0.11628458 0.10562723
#> [21,] 0.31883748 0.12754987  0.007317272 0.06433125 0.04891435 0.16862105
#> [22,] 0.22864540 0.23920396  0.012607391 0.22560623 0.21845099 0.19112124
#> [23,] 0.27940011 0.21538014  0.019596423 0.22653087 0.20941518 0.15207721
#> [24,] 0.70200782 0.66535725  0.037423706 0.63644173 0.50238427 0.42201386
#> [25,] 0.15745224 0.04471080  0.054405834 0.42278424 0.47572608 0.57947878
#>              V43        V44        V45        V46         V47         V48
#>  [1,] 0.28166548 0.19048890 0.07962270 0.02763873 0.007007626 0.045055724
#>  [2,] 0.24878587 0.24144985 0.21196234 0.20791615 0.128665064 0.081308179
#>  [3,] 0.30833650 0.32425554 0.47423138 0.35833769 0.132342994 0.135505537
#>  [4,] 0.25971056 0.25077679 0.22124922 0.18970465 0.100329472 0.047160292
#>  [5,] 0.26821997 0.14851707 0.19221792 0.18610161 0.172852028 0.147608125
#>  [6,] 0.22405589 0.23014604 0.21736701 0.22523300 0.180814951 0.113830730
#>  [7,] 0.14746234 0.11384754 0.11609129 0.12396281 0.053824589 0.003043729
#>  [8,] 0.16398480 0.08443788 0.04801037 0.05418122 0.069579184 0.040441328
#>  [9,] 0.14554845 0.25784170 0.15820578 0.07963636 0.119090097 0.108676980
#> [10,] 0.30920099 0.26705178 0.15870546 0.10897345 0.089569289 0.024622887
#> [11,] 0.21279250 0.17108874 0.12471116 0.09206066 0.104688445 0.089641482
#> [12,] 0.06656456 0.13932803 0.09395023 0.06726046 0.097325030 0.107585998
#> [13,] 0.48167728 0.42283101 0.52258378 0.47193175 0.300212228 0.191248600
#> [14,] 0.37326556 0.39456861 0.36158403 0.17427487 0.123292227 0.131966507
#> [15,] 0.13535599 0.09772252 0.08059355 0.07606172 0.040543224 0.021889903
#> [16,] 0.07173113 0.11936174 0.11092416 0.04497256 0.060884213 0.049163513
#> [17,] 0.19573650 0.14155958 0.13717491 0.14478069 0.086336056 0.070759592
#> [18,] 0.59583132 0.38852834 0.43241384 0.47388072 0.373104697 0.248123745
#> [19,] 0.15750000 0.22280000 0.15820000 0.14330000 0.163400000 0.113300000
#> [20,] 0.12873847 0.08352420 0.07679481 0.08448160 0.114332341 0.070950587
#> [21,] 0.24823396 0.27688990 0.23177171 0.16679580 0.098770975 0.101772243
#> [22,] 0.15221923 0.11278163 0.08660114 0.06951223 0.032070891 0.030378268
#> [23,] 0.16026581 0.10094227 0.05315378 0.04243348 0.040471077 0.038230082
#> [24,] 0.28340762 0.16322512 0.13084532 0.07568405 0.079149596 0.079862842
#> [25,] 0.43910269 0.24661101 0.09989312 0.05339263 0.040292598 0.074426148
#>               V49          V5         V50           V51         V52         V53
#>  [1,] 0.025062287 0.154352859 0.007148479  0.0055713096 0.021531853 0.011129862
#>  [2,] 0.045472913 0.088069253 0.020780684  0.0172554094 0.006773681 0.016282192
#>  [3,] 0.069494435 0.085441476 0.028206235  0.0223634793 0.028824174 0.015907698
#>  [4,] 0.041012188 0.120641321 0.009189207 -0.0020475292 0.013626263 0.018664011
#>  [5,] 0.070763384 0.099635729 0.012515670  0.0191664643 0.015896817 0.009699371
#>  [6,] 0.057367637 0.084568531 0.017920131  0.0191009878 0.011805255 0.005526660
#>  [7,] 0.007975319 0.136551062 0.014702525  0.0139650140 0.024634489 0.009212164
#>  [8,] 0.035423880 0.073635930 0.022413687  0.0131500024 0.033790052 0.018651794
#>  [9,] 0.083087257 0.021215534 0.011368192  0.0319414879 0.018627185 0.015698279
#> [10,] 0.034231106 0.127205407 0.015701877  0.0079810069 0.006106960 0.005910408
#> [11,] 0.057005828 0.117002659 0.022858355  0.0171452463 0.013323074 0.013571138
#> [12,] 0.062120961 0.055594593 0.023970942  0.0131695202 0.013084229 0.002624448
#> [13,] 0.108733699 0.083695686 0.041088855  0.0342916553 0.027063099 0.015615745
#> [14,] 0.072956210 0.078182194 0.018902585  0.0208394400 0.014442124 0.012451361
#> [15,] 0.002614698 0.033924541 0.015677889  0.0093954507 0.007560604 0.004984404
#> [16,] 0.023701703 0.074181133 0.009570290  0.0152626814 0.007689587 0.011154101
#> [17,] 0.052543290 0.062212051 0.030698750  0.0193484621 0.016679301 0.011221254
#> [18,] 0.162604488 0.062055257 0.061674854  0.0320445881 0.016161404 0.016017739
#> [19,] 0.056700000 0.047600000 0.013300000  0.0170000000 0.003500000 0.005200000
#> [20,] 0.022133362 0.034359141 0.025298531  0.0160998773 0.016160829 0.012496891
#> [21,] 0.059084582 0.017124470 0.016177903  0.0107199285 0.007525162 0.001754285
#> [22,] 0.019662877 0.024796918 0.007869534  0.0081411722 0.012214569 0.008338369
#> [23,] 0.014998369 0.045576421 0.009283597  0.0051789607 0.004720792 0.006294019
#> [24,] 0.035286873 0.009844977 0.022369706  0.0219020865 0.003889963 0.006393104
#> [25,] 0.028286104 0.043591714 0.005239329  0.0007644499 0.006811711 0.013788556
#>               V54         V55         V56         V57         V58         V59
#>  [1,] 0.005483208 0.020736032 0.024850009 0.007717743 0.006762244 0.013590189
#>  [2,] 0.014882417 0.019252272 0.015513324 0.010218014 0.012957366 0.012359765
#>  [3,] 0.017945672 0.014984499 0.019804435 0.013512667 0.019035540 0.018416541
#>  [4,] 0.007201935 0.006200450 0.013907124 0.007926630 0.007920009 0.013239122
#>  [5,] 0.009549851 0.008084052 0.006563789 0.008934835 0.010990636 0.008985692
#>  [6,] 0.005141841 0.004018488 0.002383889 0.003714482 0.001442627 0.002506045
#>  [7,] 0.023576002 0.006484203 0.010609627 0.009491663 0.007060057 0.009232391
#>  [8,] 0.017804488 0.028544560 0.012491021 0.002046794 0.016275504 0.015978459
#>  [9,] 0.013996322 0.025226668 0.004619831 0.005180857 0.015249947 0.009900791
#> [10,] 0.023250969 0.016090449 0.008824175 0.014016658 0.015407893 0.014525034
#> [11,] 0.014675434 0.006986477 0.010424475 0.006118760 0.010702719 0.006928661
#> [12,] 0.005665972 0.005311125 0.008338353 0.010586012 0.010869293 0.006965422
#> [13,] 0.014361751 0.013284945 0.009754724 0.008519511 0.010714609 0.011226531
#> [14,] 0.013023832 0.009987798 0.006210710 0.005136000 0.006982124 0.007906617
#> [15,] 0.011367272 0.010698323 0.007946474 0.011436021 0.008571338 0.007330725
#> [16,] 0.008870953 0.004234481 0.007293783 0.003920200 0.003226481 0.003878156
#> [17,] 0.017737582 0.013535466 0.008252753 0.011302620 0.010753354 0.002323082
#> [18,] 0.009217479 0.016845325 0.012800976 0.001113971 0.016931936 0.011111630
#> [19,] 0.008300000 0.007800000 0.007500000 0.010500000 0.016000000 0.009500000
#> [20,] 0.007135155 0.009543661 0.002111189 0.007334361 0.004769144 0.006602378
#> [21,] 0.008892636 0.006239361 0.002143394 0.002957297 0.003605675 0.007464269
#> [22,] 0.008024895 0.005283099 0.003879298 0.003350680 0.005238876 0.003649808
#> [23,] 0.006922569 0.006481405 0.005084430 0.005799176 0.005402498 0.005040113
#> [24,] 0.008540575 0.012167798 0.009305167 0.005067848 0.004897448 0.008857820
#> [25,] 0.013888340 0.010679151 0.018469929 0.010880362 0.003057970 0.006240411
#>                V6         V60         V7           V8          V9
#>  [1,] 0.081677975 0.004462269 0.07102710  0.105270439 0.095685674
#>  [2,] 0.084108021 0.004570924 0.09731498  0.090590608 0.149249208
#>  [3,] 0.147060847 0.013613532 0.18651152  0.412059414 0.625133503
#>  [4,] 0.137430923 0.004872268 0.14186951 -0.004111759 0.007543977
#>  [5,] 0.104964296 0.005712369 0.10732108  0.118015471 0.155431211
#>  [6,] 0.077317226 0.005085777 0.09548497  0.134703143 0.193440550
#>  [7,] 0.153260246 0.006535494 0.21571483  0.247071498 0.266699157
#>  [8,] 0.008784418 0.007660822 0.05338331  0.015891252 0.083281318
#>  [9,] 0.157519471 0.015128652 0.19184082  0.161661911 0.272991571
#> [10,] 0.063578676 0.010290380 0.08219694  0.039644585 0.066400393
#> [11,] 0.156883392 0.004528822 0.15021365  0.126314384 0.170334601
#> [12,] 0.119808050 0.005198092 0.12906839  0.145760717 0.175533150
#> [13,] 0.139633160 0.008578554 0.13081356  0.162557636 0.232095394
#> [14,] 0.124666334 0.005087970 0.13965062  0.203044413 0.335147527
#> [15,] 0.067139807 0.005628435 0.08408446  0.040158870 0.061243810
#> [16,] 0.091418064 0.003248851 0.12360525  0.086830807 0.114354829
#> [17,] 0.102422635 0.004406388 0.14091692  0.139764021 0.176125019
#> [18,] 0.097618887 0.010100321 0.19121743  0.223954933 0.057119939
#> [19,] 0.081600000 0.001100000 0.09930000  0.031500000 0.073600000
#> [20,] 0.047479056 0.003517129 0.04607852  0.088111630 0.119324931
#> [21,] 0.092411265 0.003848614 0.09460344  0.038026275 0.072322656
#> [22,] 0.053203121 0.003544713 0.06635881  0.089183117 0.088696070
#> [23,] 0.084005286 0.004482150 0.08506609  0.091826024 0.106852238
#> [24,] 0.060569959 0.002431835 0.04802711  0.069612943 0.002779102
#> [25,] 0.095086389 0.014331121 0.09777745  0.155343901 0.042137065
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841