Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11         V12        V13        V14
#>  [1,] 0.052209058 0.20439820 0.28031178 0.289301341 0.32620491 0.32086119
#>  [2,] 0.023783322 0.20470676 0.24914519 0.291173258 0.27568373 0.24565998
#>  [3,] 0.040330750 0.25570538 0.29702007 0.313834069 0.27094567 0.20183689
#>  [4,] 0.054961850 0.14505737 0.23588586 0.323656317 0.39799569 0.46804995
#>  [5,] 0.060245193 0.23626783 0.30337220 0.407058569 0.50104139 0.47720195
#>  [6,] 0.032917747 0.48021315 0.60732245 0.595034895 0.71584986 0.67698596
#>  [7,] 0.030863282 0.28148027 0.39144839 0.504014694 0.50017439 0.49789998
#>  [8,] 0.107214338 0.55845138 0.41779354 0.149209935 0.12366279 0.24205646
#>  [9,] 0.021984917 0.18360066 0.26994237 0.297212074 0.28144157 0.27379150
#> [10,] 0.019114132 0.21638761 0.11718672 0.009364452 0.14374996 0.25462273
#> [11,] 0.028533362 0.37834979 0.38958126 0.455144951 0.47152369 0.48172659
#> [12,] 0.030396667 0.48389715 0.56405765 0.535616827 0.49327676 0.36272168
#> [13,] 0.057632520 0.33270817 0.23752621 0.175240183 0.40153893 0.53081236
#> [14,] 0.008889321 0.06990282 0.06514084 0.058208107 0.14258294 0.13736360
#> [15,] 0.019805483 0.07680323 0.19011576 0.416970016 0.30194566 0.27428161
#> [16,] 0.031157263 0.34119641 0.17621630 0.069894407 0.06435608 0.37835656
#> [17,] 0.014054676 0.08664330 0.07590873 0.046332632 0.07640232 0.08474727
#> [18,] 0.031329292 0.10235155 0.10653625 0.244960088 0.21533353 0.12895699
#> [19,] 0.069880885 0.25286263 0.25733405 0.108177738 0.12447269 0.45954777
#> [20,] 0.021530286 0.40093609 0.31487086 0.126898945 0.24873578 0.51857905
#> [21,] 0.015710953 0.05151792 0.07031676 0.105656431 0.13422189 0.16370771
#> [22,] 0.017150328 0.07674323 0.08908594 0.084485859 0.16070017 0.22585233
#> [23,] 0.029198681 0.04378268 0.08396486 0.159861371 0.19510415 0.08662008
#> [24,] 0.003372374 0.10398195 0.12253810 0.135444493 0.19238563 0.21045792
#> [25,] 0.018925733 0.09447076 0.08291167 0.098154023 0.10004241 0.09020685
#>               V15       V16        V17       V18       V19          V2
#>  [1,] 0.344597441 0.3948660 0.36647499 0.2612451 0.3460855 0.083989675
#>  [2,] 0.202692659 0.2329568 0.25835134 0.3513192 0.5425114 0.038690973
#>  [3,] 0.150267751 0.1900249 0.19398925 0.2287997 0.2805399 0.026001135
#>  [4,] 0.568424181 0.6408527 0.70116138 0.8082867 0.9288897 0.056786135
#>  [5,] 0.590210772 0.6996780 0.90328967 0.9118243 0.8963431 0.122797047
#>  [6,] 0.687547739 0.6661628 0.82884998 0.9609810 0.9158564 0.002731238
#>  [7,] 0.563309675 0.7361405 0.82556140 0.9206835 1.0015011 0.052180733
#>  [8,] 0.449976184 0.4465160 0.27644882 0.2269694 0.3988936 0.104894605
#>  [9,] 0.218046728 0.1311589 0.18681106 0.2517342 0.2790794 0.027676685
#> [10,] 0.424645541 0.6318178 0.71701979 0.7914019 0.9285480 0.028948044
#> [11,] 0.488999123 0.5654646 0.70752048 0.7711603 0.8123566 0.085558935
#> [12,] 0.237476415 0.2587847 0.30529625 0.3584150 0.5523963 0.031430930
#> [13,] 0.219028715 0.3826035 0.55572416 0.5719637 0.7906116 0.057765309
#> [14,] 0.145289880 0.1329917 0.09374436 0.1360027 0.1644143 0.006291002
#> [15,] 0.177435424 0.1375735 0.19509149 0.4416904 0.3216792 0.069906276
#> [16,] 0.534735874 0.5453937 0.52054026 0.4371763 0.2246869 0.094084679
#> [17,] 0.007059842 0.1028760 0.16060655 0.4153742 0.3457429 0.011016125
#> [18,] 0.169616827 0.2353006 0.25858210 0.2433225 0.2347693 0.039587219
#> [19,] 0.686901232 0.6959730 0.72636655 0.5747305 0.3882656 0.043423345
#> [20,] 0.641947684 0.7464490 0.82501262 0.7286142 0.4522352 0.026814304
#> [21,] 0.208059838 0.3629557 0.52051598 0.4944644 0.3983806 0.014582176
#> [22,] 0.217157824 0.2011694 0.21273724 0.3153537 0.4572665 0.017924054
#> [23,] 0.058846058 0.1601857 0.19884109 0.3394141 0.2007466 0.004726300
#> [24,] 0.257067677 0.2877368 0.25054745 0.2365090 0.2503085 0.002319460
#> [25,] 0.123514910 0.1866079 0.26149051 0.4434134 0.5279210 0.022477567
#>             V20       V21       V22         V23         V24        V25
#>  [1,] 0.5065579 0.6203909 0.6290344  0.58337204  0.61015855 0.61139187
#>  [2,] 0.6600225 0.7670248 0.7928865  0.80014792  0.91529818 0.94741199
#>  [3,] 0.3557187 0.4770570 0.6150649  0.72579441  0.78811630 0.89466368
#>  [4,] 0.9166338 0.9015561 0.9035051  0.91749731  0.85543546 0.67767336
#>  [5,] 0.7886113 0.8051013 0.6024749  0.34350219  0.30650489 0.20649726
#>  [6,] 0.8433040 0.6499600 0.2768937 -0.08309940 -0.06995962 0.25600744
#>  [7,] 0.9974112 0.8280338 0.6499931  0.59534631  0.48209097 0.31126581
#>  [8,] 0.8516161 0.9511747 0.4315448  0.08937783  0.17951471 0.17284310
#>  [9,] 0.3542439 0.4392885 0.4363447  0.50099747  0.58765272 0.65499191
#> [10,] 0.9597975 0.9785805 0.9820074  0.88702712  0.59376975 0.36540370
#> [11,] 0.8780539 0.8422805 0.8934632  0.99651745  1.04249428 0.95328357
#> [12,] 0.7044695 0.8198151 0.8533492  0.81219846  0.83719084 0.83782704
#> [13,] 0.8588371 0.5666590 0.6672151  0.84259395  0.68788702 0.76585506
#> [14,] 0.2061579 0.2865377 0.4645554  0.60398192  0.71737111 0.72994409
#> [15,] 0.6721991 0.7856717 0.7567600  0.74682679  0.80627075 1.00933585
#> [16,] 0.4286536 0.7622094 0.9588548  0.93230914  0.56543914 0.52902305
#> [17,] 0.3014334 0.1970398 0.1936630  0.17984122  0.19849355 0.01364755
#> [18,] 0.2003704 0.2366422 0.1049966  0.19237572  0.19395383 0.11653979
#> [19,] 0.2741237 0.5261938 0.7666880  0.86319401  0.94677800 0.91566858
#> [20,] 0.3523499 0.4367320 0.6338195  0.79057474  0.92035582 0.76953217
#> [21,] 0.3853028 0.4808830 0.5219858  0.48507477  0.54815348 0.63543587
#> [22,] 0.5709307 0.5893307 0.6408861  0.65575225  0.65930714 0.78476144
#> [23,] 0.1922842 0.5665829 0.6466507  0.69858273  0.82735798 0.74960503
#> [24,] 0.3171966 0.4150930 0.4624829  0.55983262  0.74106249 0.87326091
#> [25,] 0.5574122 0.6146519 0.6056526  0.65667805  0.68036067 0.51573998
#>             V26        V27        V28       V29           V3       V30
#>  [1,] 0.6944050 0.75535047 0.82066457 0.7999826  0.094297793 0.8290643
#>  [2,] 0.9823818 0.99919145 0.97156655 0.8563918  0.037096539 0.6457253
#>  [3,] 0.9378543 0.93920733 0.93003679 0.9036142  0.032218089 0.8255936
#>  [4,] 0.5757611 0.55283582 0.54078445 0.4874330  0.068452434 0.3813624
#>  [5,] 0.1194942 0.15386475 0.17424127 0.1494023  0.119743019 0.1367539
#>  [6,] 0.4902159 0.51805024 0.47250070 0.2547388  0.075392150 0.1172752
#>  [7,] 0.1495875 0.09436786 0.02531468 0.1012146  0.065690457 0.2664723
#>  [8,] 0.3105793 0.19585705 0.36724567 0.4963789  0.066777127 0.6631984
#>  [9,] 0.7787528 0.90519246 0.95281244 0.8658474  0.037012348 0.6658498
#> [10,] 0.3264253 0.44297720 0.63962324 0.7099903  0.028331924 0.5929248
#> [11,] 0.8883058 0.86990579 0.74448422 0.6860198  0.100251950 0.5673507
#> [12,] 0.8631698 0.90314240 0.86080604 0.6413876  0.040138321 0.4216925
#> [13,] 0.8267182 0.74470010 0.65384745 0.3852445  0.065236453 0.6563197
#> [14,] 0.7509985 0.72815024 0.83214352 0.9903402  0.019189031 0.9182741
#> [15,] 0.9442655 0.71440086 0.51550179 0.4179179  0.079085810 0.3777465
#> [16,] 0.6740725 0.56464822 0.52999589 0.2113565  0.128005593 0.2082308
#> [17,] 0.1593680 0.40582599 0.64054536 0.7914622 -0.009867996 0.7236736
#> [18,] 0.2228423 0.41329048 0.64769845 0.7130601  0.037629706 0.6832051
#> [19,] 0.6376424 0.62468443 0.65824493 0.4982288  0.051636548 0.5639929
#> [20,] 0.6850927 0.79278440 0.66766454 0.4637899  0.053842977 0.4909368
#> [21,] 0.6946249 0.79810184 0.88070547 0.9143209  0.017071324 0.8856438
#> [22,] 0.9046189 0.98958096 0.91522338 0.7549865  0.013976732 0.6487025
#> [23,] 0.6165587 0.49630878 0.32527745 0.4500195  0.008126466 0.5033996
#> [24,] 0.9110528 0.83714115 0.76869034 0.6476760 -0.001449046 0.5068567
#> [25,] 0.4785614 0.58104540 0.77378997 0.8320785  0.014711284 0.7311413
#>             V31       V32       V33       V34       V35          V36        V37
#>  [1,] 0.8517209 0.7495313 0.4904594 0.4164634 0.3933495  0.241169435 0.29366457
#>  [2,] 0.4429267 0.3343406 0.2854288 0.2060783 0.1462491  0.069654212 0.09622316
#>  [3,] 0.7296095 0.5245415 0.4459305 0.2958484 0.1553419  0.118500332 0.14241805
#>  [4,] 0.2603420 0.1753429 0.2078362 0.2997832 0.3933188  0.363066097 0.32257128
#>  [5,] 0.2010561 0.2644780 0.3414404 0.4282816 0.5277116  0.218348149 0.13212023
#>  [6,] 0.3163721 0.3365161 0.3492065 0.3374753 0.4335593 -0.005705181 0.12247405
#>  [7,] 0.2706466 0.2907808 0.2881681 0.4001024 0.3676541  0.313509823 0.45487560
#>  [8,] 0.3129838 0.3530785 0.5766503 0.7096581 0.8545639  0.835127996 0.79604926
#>  [9,] 0.5826950 0.4447063 0.3402247 0.2394391 0.1276149  0.073963512 0.18487325
#> [10,] 0.3313537 0.3469104 0.4185143 0.3036438 0.2216169  0.141611959 0.09502015
#> [11,] 0.3252403 0.1889342 0.1339906 0.2357459 0.3015220  0.207521845 0.23474775
#> [12,] 0.3387420 0.3734511 0.3808780 0.3607349 0.1982021  0.316513829 0.31878891
#> [13,] 0.7285282 0.8236487 0.7185853 0.5931400 0.3560358  0.121038906 0.10141370
#> [14,] 0.7731081 0.7147969 0.7614119 0.6622283 0.5416724  0.437255882 0.28963038
#> [15,] 0.5441073 0.5653358 0.6444170 0.5638900 0.5556483  0.545769142 0.48162378
#> [16,] 0.5637541 0.6243543 0.3095734 0.1956017 0.2948419  0.518939872 0.62701877
#> [17,] 0.5079437 0.1429858 0.4573291 0.4583514 0.2198260  0.351421254 0.48395547
#> [18,] 0.7161705 0.6263110 0.6590948 0.8079304 0.8835736  0.902504091 0.79245544
#> [19,] 0.4633283 0.2209865 0.2517564 0.1665440 0.4990147  0.734734719 0.76676959
#> [20,] 0.5328897 0.3427270 0.2271127 0.1827661 0.5240208  0.655260364 0.59616590
#> [21,] 0.8214753 0.7120734 0.6934598 0.6637546 0.6862861  0.683805604 0.62364456
#> [22,] 0.6015277 0.4628573 0.3818528 0.2655776 0.2135708  0.289701910 0.29852216
#> [23,] 0.5455250 0.5458975 0.7064084 0.6046670 0.5780057  0.576909532 0.55951126
#> [24,] 0.4293157 0.3246672 0.2222726 0.2378154 0.2730958  0.282175017 0.19557948
#> [25,] 0.7268384 0.6791440 0.5510031 0.5023728 0.5327916  0.573228106 0.45081178
#>             V38         V39           V4       V40         V41        V42
#>  [1,] 0.3326279  0.35392677  0.087117345 0.2950007 0.564740598 0.66231520
#>  [2,] 0.1411556  0.13466192  0.030222334 0.1410873 0.069095189 0.13146329
#>  [3,] 0.1594854  0.18513872  0.043485584 0.1242751 0.162554656 0.18706960
#>  [4,] 0.3174001  0.39139834  0.083050847 0.3626851 0.323979180 0.35773437
#>  [5,] 0.2812357  0.31827813  0.163670065 0.1669857 0.070806532 0.07668456
#>  [6,] 0.1489647  0.14042112  0.142451783 0.1235821 0.005080317 0.06915582
#>  [7,] 0.4316709  0.17951527  0.051740779 0.3379450 0.620350504 0.44205970
#>  [8,] 0.8808175  0.81583646  0.062973750 0.5748915 0.288677580 0.16399989
#>  [9,] 0.1499374  0.21853861  0.063087690 0.2231555 0.241633251 0.33826269
#> [10,] 0.1618183  0.12684736  0.056385196 0.1342260 0.203493237 0.20143645
#> [11,] 0.3233099  0.29985639  0.054447201 0.2494964 0.360466885 0.34227861
#> [12,] 0.3334459  0.45659890  0.046286743 0.3761237 0.160690840 0.24489208
#> [13,] 0.3891642  0.60798887  0.101398493 0.4247916 0.451565733 0.49043421
#> [14,] 0.2609911  0.29278362  0.028514369 0.2620270 0.162832914 0.10138489
#> [15,] 0.1435292 -0.06346784  0.067526172 0.3232353 0.399680131 0.60136351
#> [16,] 0.4462147  0.53961132  0.138047239 0.6002743 0.506384904 0.25507192
#> [17,] 0.5882606  0.67414578  0.004413569 0.5280689 0.202871237 0.20220946
#> [18,] 0.7363313  0.70433204  0.026372682 0.6955978 0.625007675 0.55057396
#> [19,] 0.4208686  0.12213936  0.034913129 0.3315511 0.441564086 0.41294044
#> [20,] 0.3088585  0.26610594  0.026854327 0.4167471 0.360410930 0.32406400
#> [21,] 0.4906056  0.29135018  0.007763174 0.1551236 0.124525732 0.15860541
#> [22,] 0.2812309  0.23950727  0.023192365 0.1629845 0.136657057 0.19156069
#> [23,] 0.4810471  0.48474112  0.001029793 0.6894815 0.772628895 0.71131164
#> [24,] 0.1833238  0.19874157 -0.008448697 0.1773280 0.176274003 0.15530272
#> [25,] 0.3643383  0.33757474  0.021009137 0.2770360 0.173393410 0.14037729
#>              V43        V44         V45         V46        V47        V48
#>  [1,] 0.50501954 0.39460322  0.60103192  0.53550626 0.29639917 0.17347696
#>  [2,] 0.21412001 0.23442703  0.20627430  0.19674269 0.15428069 0.09577630
#>  [3,] 0.19647813 0.14471048  0.10106700  0.06390949 0.09458344 0.09856909
#>  [4,] 0.35085628 0.25022855  0.13669690  0.08106638 0.08013917 0.04990539
#>  [5,] 0.10101912 0.04199667  0.06405721  0.07763862 0.10775698 0.11491694
#>  [6,] 0.11127336 0.06324764  0.07300805  0.04128771 0.03941278 0.05832586
#>  [7,] 0.18509881 0.14200142  0.05794728  0.05658907 0.06703246 0.02610489
#>  [8,] 0.17655373 0.27904699  0.48794342  0.38395044 0.18905958 0.15589434
#>  [9,] 0.22074054 0.10208386  0.10918768  0.18323914 0.19342498 0.14006046
#> [10,] 0.12256485 0.19035368  0.16736712  0.10529622 0.09607756 0.06750275
#> [11,] 0.09927629 0.16443292  0.20248054  0.15827717 0.09424853 0.07713498
#> [12,] 0.37332083 0.41732723  0.34947392  0.16904149 0.11208476 0.11213676
#> [13,] 0.42875041 0.37778408  0.50228907  0.51926012 0.30035815 0.23286504
#> [14,] 0.14423895 0.11231958  0.07540129  0.03625950 0.02183631 0.02892832
#> [15,] 0.47610362 0.26395130  0.12063403  0.06870433 0.05802824 0.07095649
#> [16,] 0.19896510 0.17346717  0.13020896  0.06270025 0.11094981 0.10375347
#> [17,] 0.16005973 0.04639967 -0.01806931 -0.03598942 0.05643394 0.06943095
#> [18,] 0.39770583 0.24057716  0.17547142  0.13102146 0.13559509 0.11892362
#> [19,] 0.29095159 0.21797007  0.25394639  0.25833917 0.22903435 0.19456989
#> [20,] 0.24450721 0.16080295  0.16335327  0.14355506 0.12377509 0.10535196
#> [21,] 0.19458325 0.16636827  0.15710325  0.14106642 0.08221446 0.06542326
#> [22,] 0.12487675 0.06592823  0.09528420  0.07141824 0.08260049 0.06663850
#> [23,] 0.74736995 0.60989357  0.38268949  0.26134405 0.17059839 0.11801403
#> [24,] 0.15593003 0.12213315  0.07538082  0.07622002 0.05307716 0.03260369
#> [25,] 0.12506770 0.13883593  0.09706289  0.06574102 0.06914786 0.06976408
#>               V49          V5         V50         V51           V52         V53
#>  [1,] 0.109387139 0.116869998 0.041883729 0.026771754  0.0195767754 0.017488301
#>  [2,] 0.048546756 0.058019220 0.015115411 0.019040933  0.0119942573 0.006788546
#>  [3,] 0.056681987 0.046691783 0.014276116 0.009122615  0.0105406204 0.005182542
#>  [4,] 0.037359708 0.110347330 0.013707517 0.018999998  0.0137866398 0.013793807
#>  [5,] 0.081461693 0.173010282 0.035962118 0.024640467  0.0384585830 0.014185063
#>  [6,] 0.038442683 0.023711148 0.021942964 0.036016704  0.0300224956 0.028956021
#>  [7,] 0.024109543 0.059996781 0.018963084 0.005878306  0.0320858016 0.018982351
#>  [8,] 0.091695877 0.050268497 0.044921894 0.021121921  0.0133826679 0.011393302
#>  [9,] 0.059809554 0.073702440 0.015518152 0.019440053  0.0106365109 0.007901770
#> [10,] 0.049085000 0.092523555 0.013975734 0.007018308  0.0116215377 0.004210856
#> [11,] 0.035726079 0.144160937 0.020666224 0.019193157  0.0216181760 0.023412867
#> [12,] 0.065537755 0.080652139 0.020041550 0.020495124  0.0119328399 0.011862664
#> [13,] 0.128634423 0.220831894 0.077247855 0.020882942  0.0188959780 0.030800396
#> [14,] 0.005031214 0.025144525 0.007776889 0.008246059  0.0084315873 0.011655729
#> [15,] 0.024577192 0.136509799 0.007079693 0.003771295  0.0009912367 0.014267961
#> [16,] 0.050756445 0.166158218 0.016147488 0.020216413  0.0206644463 0.024291998
#> [17,] 0.013965298 0.039973628 0.010970663 0.005773124  0.0225572039 0.029577495
#> [18,] 0.069268150 0.044489852 0.026545214 0.014619752  0.0082879852 0.016432905
#> [19,] 0.133541264 0.035299701 0.036741264 0.013152974  0.0153410029 0.011329181
#> [20,] 0.067397274 0.023675451 0.018451091 0.018577555  0.0087224798 0.003106514
#> [21,] 0.033953499 0.009665102 0.019848805 0.010233228  0.0088672231 0.001865867
#> [22,] 0.028715707 0.038597239 0.019432745 0.006217178  0.0083130652 0.009388055
#> [23,] 0.016588263 0.030747885 0.013671936 0.007347685 -0.0045363960 0.005992978
#> [24,] 0.021060800 0.019594612 0.011306026 0.007493967  0.0106901606 0.004597526
#> [25,] 0.040027106 0.017032383 0.021809855 0.008099732  0.0057132119 0.003872989
#>                V54          V55         V56           V57         V58
#>  [1,] 0.0117467847  0.011765215 0.010162389  0.0085341964 0.010096708
#>  [2,] 0.0058397032  0.004642090 0.003195666  0.0035820748 0.002306809
#>  [3,] 0.0071715113  0.005727304 0.005485695  0.0079549530 0.006913558
#>  [4,] 0.0229894728  0.009219955 0.012378647  0.0105692396 0.013553609
#>  [5,] 0.0110130571  0.018576988 0.009746530  0.0019399166 0.011877115
#>  [6,] 0.0108636884  0.019370469 0.021218566  0.0013953459 0.048170994
#>  [7,] 0.0137241374  0.027939954 0.015381740  0.0052099389 0.012219267
#>  [8,] 0.0163610859  0.018937023 0.012585081  0.0157498232 0.013704479
#>  [9,] 0.0067880048  0.005768157 0.005343036  0.0039151576 0.005826107
#> [10,] 0.0007642387  0.000830841 0.004357392  0.0035390054 0.008952600
#> [11,] 0.0293461421  0.015694261 0.012546096  0.0106342593 0.009797557
#> [12,] 0.0123695865  0.008684733 0.007481152  0.0056071259 0.006658825
#> [13,] 0.0206094994  0.038941624 0.033483427  0.0312128574 0.038747855
#> [14,] 0.0062104457  0.005074070 0.007046832  0.0057893718 0.003282689
#> [15,] 0.0138148997  0.014231198 0.014869188  0.0140368861 0.008671248
#> [16,] 0.0130522392  0.007164992 0.013305024  0.0089959309 0.013839759
#> [17,] 0.0165910659  0.025909808 0.002301487 -0.0008245577 0.003917629
#> [18,] 0.0124264803  0.008605463 0.007717034  0.0072327757 0.011301403
#> [19,] 0.0087000747  0.003535279 0.005241190  0.0218466898 0.009141078
#> [20,] 0.0116465895  0.010561517 0.002732143  0.0103366588 0.005368062
#> [21,] 0.0086338370  0.009060008 0.002664339  0.0066409068 0.006246311
#> [22,] 0.0075693580  0.005175194 0.006705911  0.0052285127 0.005156755
#> [23,] 0.0030303914 -0.003211875 0.010638493  0.0078472161 0.006622239
#> [24,] 0.0052359424  0.004521817 0.005462641  0.0029434720 0.003598027
#> [25,] 0.0071402770  0.008910931 0.007380627  0.0109327899 0.008162560
#>               V59         V6         V60         V7         V8         V9
#>  [1,] 0.010585194 0.13250396 0.011737016 0.10952450 0.14181686 0.21663694
#>  [2,] 0.003624338 0.05610920 0.003341142 0.09031499 0.10967697 0.16018880
#>  [3,] 0.004422586 0.09893397 0.004352834 0.11232430 0.12883293 0.17324201
#>  [4,] 0.016281847 0.13939088 0.012195374 0.14358186 0.12897228 0.10713013
#>  [5,] 0.025678080 0.05936805 0.013800783 0.08231235 0.02847486 0.10499680
#>  [6,] 0.014890195 0.10392883 0.018710949 0.20580200 0.13316935 0.31679013
#>  [7,] 0.012417038 0.05737060 0.005471603 0.06663772 0.06703183 0.14660059
#>  [8,] 0.008313628 0.11095385 0.009213225 0.21927039 0.31973449 0.48867736
#>  [9,] 0.004841028 0.08077151 0.002629945 0.08229150 0.09148056 0.13683175
#> [10,] 0.006195861 0.14408619 0.001177257 0.16991410 0.16084095 0.19477324
#> [11,] 0.002766240 0.20690838 0.004729182 0.17439913 0.21802954 0.33477259
#> [12,] 0.008054527 0.14530681 0.005967976 0.16129987 0.21439140 0.35094671
#> [13,] 0.021768846 0.12865829 0.008771719 0.17479589 0.10673076 0.24571525
#> [14,] 0.008993907 0.03109745 0.005317439 0.03472818 0.03877215 0.03861486
#> [15,] 0.003864281 0.21741336 0.016318269 0.15493455 0.15802810 0.01607806
#> [16,] 0.003929984 0.16708795 0.010178427 0.07239227 0.13496788 0.20375838
#> [17,] 0.008482023 0.04689320 0.009150942 0.02776023 0.05817141 0.14348940
#> [18,] 0.012000883 0.07468180 0.008451012 0.08209544 0.09639852 0.07775104
#> [19,] 0.004429480 0.05231755 0.003394089 0.19715137 0.28422103 0.26101553
#> [20,] 0.010474984 0.11527857 0.007874072 0.10522730 0.22008047 0.31508852
#> [21,] 0.007612025 0.08118016 0.006386753 0.09596877 0.04524039 0.04080820
#> [22,] 0.003082018 0.04653144 0.004084925 0.08090520 0.06552998 0.09771303
#> [23,] 0.011733577 0.05543220 0.008246856 0.09036372 0.08815520 0.06441788
#> [24,] 0.002777124 0.08060750 0.001605806 0.09309748 0.10231387 0.09579729
#> [25,] 0.004867169 0.06356624 0.004058486 0.08900428 0.12154602 0.11478076
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696