Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.034468916 0.27879640 0.32407193 0.32698520 0.24924401 0.08924588
#>  [2,] 0.035398394 0.26658175 0.24750956 0.27823884 0.32311285 0.30547582
#>  [3,] 0.008286290 0.34350341 0.31973983 0.32006872 0.23286995 0.14276191
#>  [4,] 0.020835890 0.35646208 0.36187939 0.20136648 0.46146709 0.86202482
#>  [5,] 0.024449689 0.30330191 0.24565294 0.27827234 0.25530951 0.20432334
#>  [6,] 0.011474968 0.20093755 0.29464641 0.30252811 0.30014221 0.25007939
#>  [7,] 0.026614567 0.27138622 0.30631996 0.35750691 0.30047811 0.26706486
#>  [8,] 0.034604279 0.38086494 0.43877699 0.47168187 0.47630003 0.40215670
#>  [9,] 0.049893324 0.19017943 0.25652019 0.31453479 0.37568806 0.45396944
#> [10,] 0.021319386 0.16710351 0.19038790 0.28392739 0.32184892 0.28336025
#> [11,] 0.042026371 0.27273257 0.32005855 0.34626539 0.31579317 0.36570089
#> [12,] 0.024052457 0.36536420 0.48728076 0.47146597 0.36766937 0.30245236
#> [13,] 0.103735046 0.58334954 0.47785802 0.20220916 0.19702558 0.28312242
#> [14,] 0.020127892 0.06970960 0.10813608 0.27141534 0.31674917 0.27111148
#> [15,] 0.011980915 0.14023394 0.19151043 0.17326824 0.16589931 0.17827991
#> [16,] 0.007373420 0.08192897 0.11650216 0.07252379 0.10566626 0.14155944
#> [17,] 0.029375082 0.12160596 0.16960269 0.24768168 0.30304142 0.41286273
#> [18,] 0.025148815 0.36515449 0.29646720 0.10878999 0.21991709 0.51860842
#> [19,] 0.044428563 0.24682201 0.25316838 0.33168729 0.30329571 0.25320421
#> [20,] 0.011757476 0.07947622 0.20912453 0.38143117 0.26933776 0.21820360
#> [21,] 0.029695136 0.23806330 0.35812537 0.37893341 0.39210094 0.41281674
#> [22,] 0.007815088 0.06078785 0.05763436 0.09829497 0.13708321 0.10069458
#> [23,] 0.015116723 0.10437065 0.08387696 0.20459357 0.15452413 0.12529943
#> [24,] 0.015217365 0.13339486 0.11535270 0.10955511 0.09863568 0.12062555
#> [25,] 0.012769732 0.06633248 0.06750073 0.07291228 0.08592629 0.12874212
#>              V15        V16       V17       V18       V19           V2
#>  [1,] 0.12603585 0.18186779 0.2125256 0.2520630 0.2280149 0.0424996401
#>  [2,] 0.31086488 0.33918391 0.2986863 0.1948060 0.3216899 0.0504568609
#>  [3,] 0.18970434 0.21390632 0.1801915 0.1739477 0.1336617 0.0143110563
#>  [4,] 0.81185700 0.73927980 0.6828578 0.7695565 0.8680556 0.0319982388
#>  [5,] 0.11603037 0.17740772 0.1854272 0.2458123 0.2129825 0.0307979261
#>  [6,] 0.14353649 0.06921854 0.1790346 0.2389370 0.2676309 0.0236502617
#>  [7,] 0.25626312 0.29081989 0.3417549 0.4347326 0.5499682 0.0317818655
#>  [8,] 0.28164418 0.32338998 0.3730066 0.3875612 0.5415768 0.0351012450
#>  [9,] 0.57732427 0.66402032 0.7631962 0.8196511 0.8776220 0.0643512235
#> [10,] 0.21931526 0.17457311 0.1420189 0.1826955 0.3415932 0.0495181849
#> [11,] 0.27699517 0.15004790 0.1828701 0.2183790 0.3332708 0.0388358132
#> [12,] 0.29190050 0.38730777 0.3653071 0.3390211 0.3808784 0.0367935577
#> [13,] 0.55862741 0.56872677 0.2228998 0.1670131 0.4095210 0.1000497848
#> [14,] 0.03279144 0.09079385 0.1807490 0.4392869 0.4940883 0.0202844172
#> [15,] 0.27669877 0.31090333 0.2726911 0.2221031 0.2750874 0.0223351774
#> [16,] 0.17737888 0.22653469 0.2337559 0.2343757 0.1855876 0.0001935985
#> [17,] 0.59236545 0.71975839 0.8159845 0.8251334 0.7942530 0.0306463177
#> [18,] 0.67095983 0.76116766 0.8296286 0.7056935 0.4369068 0.0327979987
#> [19,] 0.33433890 0.36837220 0.4234151 0.5324252 0.6911999 0.0714160013
#> [20,] 0.11079899 0.10718566 0.1847247 0.4141949 0.3096346 0.0602041394
#> [21,] 0.57285784 0.73979021 0.9069818 0.9515199 0.9522577 0.0335541431
#> [22,] 0.04837218 0.07269543 0.1043355 0.1597687 0.1757820 0.0054058133
#> [23,] 0.15753319 0.17467079 0.1626719 0.2918097 0.2269086 0.0190481171
#> [24,] 0.10489932 0.19589375 0.3432299 0.3453998 0.3390343 0.0130779115
#> [25,] 0.14713936 0.22860179 0.2700414 0.3148700 0.3570651 0.0146198794
#>              V20        V21       V22       V23       V24        V25       V26
#>  [1,] 0.22069369 0.38162390 0.6687471 0.8485764 0.8757504 0.89662437 0.9222313
#>  [2,] 0.43116091 0.49835575 0.4668411 0.4746458 0.5161428 0.54917143 0.6253692
#>  [3,] 0.08713566 0.12690936 0.2918571 0.5009829 0.6504681 0.76863262 0.8867171
#>  [4,] 0.86559088 0.91240508 0.9610437 0.9675446 0.8840323 0.80155031 0.8169705
#>  [5,] 0.16318259 0.07422105 0.1711456 0.3603167 0.5199790 0.63821030 0.7553756
#>  [6,] 0.31002354 0.36173958 0.3955806 0.4953245 0.6373586 0.71041502 0.7889593
#>  [7,] 0.67996322 0.84523446 0.9128126 0.8881631 0.9838035 0.99462642 0.9915963
#>  [8,] 0.69385486 0.71200069 0.7662359 0.7653016 0.7669769 0.84105018 0.9356815
#>  [9,] 0.88208247 0.86860198 0.8450214 0.7704958 0.6507547 0.51267601 0.4307998
#> [10,] 0.35910298 0.43828135 0.5925489 0.6917977 0.7372540 0.75890991 0.7779549
#> [11,] 0.45203069 0.66965987 0.7156283 0.7610677 0.7802288 0.78515613 0.6553063
#> [12,] 0.50283230 0.72482076 0.9286968 0.9766389 0.9452294 0.87767232 0.7922177
#> [13,] 0.86530904 0.88185404 0.3911913 0.1332678 0.1549783 0.20465702 0.3133812
#> [14,] 0.68011137 0.72461459 0.7309476 0.7169814 0.6576442 0.76800912 0.7521709
#> [15,] 0.38486163 0.32198113 0.4302271 0.5747067 0.6265269 0.69538227 0.7268905
#> [16,] 0.15871402 0.21916550 0.3176509 0.4540841 0.6421897 0.79511644 0.8890688
#> [17,] 0.79245114 0.83637339 0.8717769 0.8091922 0.8370423 0.86168451 0.8751034
#> [18,] 0.33613751 0.48429785 0.7029845 0.8331363 0.9560185 0.83140107 0.7050109
#> [19,] 0.82335329 0.88605371 0.8197756 0.7504947 0.7308580 0.75400882 0.8067822
#> [20,] 0.62178221 0.74696347 0.7628881 0.7826409 0.8262173 1.00923309 0.9501673
#> [21,] 0.85484886 0.82619002 0.8061160 0.5516434 0.4564955 0.61070743 0.6554612
#> [22,] 0.17586896 0.23250274 0.3968514 0.4848669 0.6145876 0.71424193 0.7793074
#> [23,] 0.26234638 0.29938874 0.1881947 0.1622998 0.1767166 0.08765017 0.2312882
#> [24,] 0.45299325 0.58218292 0.7437727 0.8518991 0.8344051 0.84908180 0.9348508
#> [25,] 0.36617515 0.39638687 0.4015968 0.3949146 0.4729730 0.46700724 0.4995328
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.9268242 0.9237806 0.8184887 0.045625205 0.6788281 0.6777984 0.5654062
#>  [2,] 0.6639177 0.7866705 0.8274296 0.066843069 0.8473534 0.8291735 0.6987800
#>  [3,] 0.9501825 0.9615899 0.9121535 0.016545029 0.9235417 0.8862269 0.8061640
#>  [4,] 0.6644379 0.4992979 0.4987026 0.026353774 0.3614831 0.2593906 0.1000796
#>  [5,] 0.8656400 0.9006893 0.8798523 0.022660918 0.9292611 0.9333065 0.8557290
#>  [6,] 0.9586326 1.0778280 0.9533996 0.019112648 0.6778559 0.6246166 0.4838066
#>  [7,] 0.9782200 0.9694582 0.8526112 0.032910960 0.6108474 0.4592453 0.3630255
#>  [8,] 0.9571791 0.8055256 0.5775429 0.040618054 0.4636945 0.3906918 0.4494947
#>  [9,] 0.3985625 0.3547386 0.3185785 0.073874967 0.3333047 0.2502593 0.2630536
#> [10,] 0.9354719 0.9923178 0.8776590 0.056439052 0.7157397 0.6689510 0.6242449
#> [11,] 0.7403234 0.7803251 0.9850272 0.044761314 1.0335044 0.8291621 0.5754767
#> [12,] 0.7049087 0.6124570 0.4444321 0.040425171 0.3287923 0.2678939 0.3169127
#> [13,] 0.2216517 0.5092724 0.5484331 0.041743977 0.6817089 0.3533684 0.3933487
#> [14,] 0.6844216 0.5528089 0.4632737 0.045937755 0.3675359 0.4093474 0.4469678
#> [15,] 0.6715715 0.7592848 0.7463060 0.022684163 0.6095266 0.4862896 0.4166371
#> [16,] 0.8421512 0.7957289 0.6797360 0.004727004 0.5413841 0.4493114 0.3641533
#> [17,] 0.7541954 0.6151059 0.6143376 0.032890817 0.6157683 0.5164711 0.3294540
#> [18,] 0.7679541 0.6451771 0.4561122 0.053644592 0.5026988 0.5101698 0.2898824
#> [19,] 0.7383614 0.5850097 0.3141811 0.049402697 0.0981365 0.2856756 0.7069497
#> [20,] 0.7470873 0.5884213 0.5359323 0.066332911 0.5382028 0.5776044 0.5143442
#> [21,] 0.6418524 0.5531606 0.5760129 0.045901181 0.5721899 0.5612362 0.3967340
#> [22,] 0.7713693 0.8507149 0.9698491 0.014861659 0.9317185 0.7921059 0.7618392
#> [23,] 0.3498182 0.4317318 0.5064332 0.032307504 0.5699760 0.6679010 0.6762792
#> [24,] 0.9772941 0.9218596 0.8541376 0.017691310 0.7671647 0.6931024 0.6326370
#> [25,] 0.6303609 0.8373321 0.9719697 0.014607076 0.9404098 0.8270910 0.6982435
#>              V33       V34        V35         V36       V37        V38
#>  [1,] 0.42760523 0.1839969 0.08545169 0.082180820 0.1584271 0.14033287
#>  [2,] 0.50612172 0.3591573 0.24545890 0.184856318 0.2344270 0.31362925
#>  [3,] 0.71647512 0.6655445 0.64270923 0.681103326 0.6201093 0.48619107
#>  [4,] 0.06442437 0.1218636 0.16627228 0.233288337 0.1685629 0.24176274
#>  [5,] 0.72157335 0.6355860 0.66042669 0.718351014 0.7445393 0.63942921
#>  [6,] 0.34388885 0.2244891 0.14407659 0.039753119 0.1934992 0.11794838
#>  [7,] 0.29166655 0.2126146 0.12784270 0.008514022 0.1153469 0.24703341
#>  [8,] 0.39307548 0.2989186 0.11559153 0.274068903 0.3051157 0.34748208
#>  [9,] 0.31677106 0.2995544 0.30011293 0.259949496 0.2821217 0.33951711
#> [10,] 0.52857238 0.4930603 0.24270523 0.078730050 0.2036054 0.18087906
#> [11,] 0.59699636 0.5410335 0.29230100 0.259293058 0.2269547 0.08198827
#> [12,] 0.39314029 0.4043702 0.26424235 0.238010431 0.3138814 0.32532412
#> [13,] 0.67147738 0.7929615 0.90505910 0.857372220 0.8542563 0.88914159
#> [14,] 0.47251046 0.4362076 0.47909503 0.484934482 0.4240886 0.34187004
#> [15,] 0.55391456 0.8452658 0.85383288 0.718443874 0.4585475 0.32883016
#> [16,] 0.34218986 0.3601352 0.36787638 0.295996943 0.1222464 0.07243885
#> [17,] 0.32363158 0.3642422 0.23720182 0.108762271 0.1870376 0.16491872
#> [18,] 0.22695390 0.1629814 0.51435790 0.691046775 0.6602626 0.33970085
#> [19,] 0.77854349 0.7927181 0.75702794 0.761542799 0.6937803 0.25706105
#> [20,] 0.60376693 0.5492087 0.57897222 0.587126143 0.4883981 0.26400032
#> [21,] 0.35923229 0.3333118 0.23346104 0.176626211 0.2493786 0.18109975
#> [22,] 0.82285022 0.6923086 0.57148405 0.469427103 0.3649985 0.37183874
#> [23,] 0.67406275 0.7767099 0.86723554 0.817392866 0.8098134 0.80600110
#> [24,] 0.59356690 0.6208459 0.65783265 0.621217411 0.5641919 0.46052955
#> [25,] 0.67086155 0.6136416 0.57183389 0.567649801 0.5309795 0.37102525
#>              V39           V4        V40       V41        V42       V43
#>  [1,] 0.13508258 0.0528917978 0.09131243 0.1384099 0.14085674 0.1905681
#>  [2,] 0.38412448 0.0526586147 0.39563585 0.5590506 0.56739351 0.4966382
#>  [3,] 0.39492889 0.0595653555 0.42722632 0.4541978 0.43646899 0.4381004
#>  [4,] 0.36865265 0.0736206584 0.33897163 0.2390676 0.15675248 0.2309571
#>  [5,] 0.57215121 0.0347098563 0.58963238 0.5256878 0.44283918 0.4734927
#>  [6,] 0.17952646 0.0614685043 0.22122724 0.2552347 0.36064114 0.2108414
#>  [7,] 0.20323203 0.0491550934 0.20706947 0.1152919 0.15433044 0.2284307
#>  [8,] 0.42257757 0.0492952851 0.33582696 0.1767507 0.31298956 0.3881541
#>  [9,] 0.30742001 0.0963917729 0.25647795 0.2606075 0.28549314 0.2236280
#> [10,] 0.30968816 0.0121420152 0.48483655 0.3913061 0.18514052 0.1144334
#> [11,] 0.22865816 0.0587068738 0.33181636 0.3555818 0.42454127 0.3464497
#> [12,] 0.49426334 0.0474718582 0.48084628 0.3240237 0.14330835 0.3003957
#> [13,] 0.75218432 0.0843826786 0.52322896 0.2635390 0.21515765 0.2027828
#> [14,] 0.27502964 0.0380449064 0.41060223 0.3443876 0.36381393 0.3433325
#> [15,] 0.26544518 0.0355403602 0.34912576 0.2826469 0.34466619 0.2998901
#> [16,] 0.12254083 0.0009558522 0.15793254 0.1812535 0.19110395 0.2076420
#> [17,] 0.12097328 0.0247131799 0.16347265 0.1172911 0.08300285 0.1018168
#> [18,] 0.20105381 0.0220715760 0.37833387 0.3800151 0.35169195 0.2578285
#> [19,] 0.33304740 0.0718984827 0.50216082 0.5051182 0.45216000 0.3891640
#> [20,] 0.01881485 0.0611890758 0.27817111 0.3438587 0.52924185 0.4432958
#> [21,] 0.24758531 0.0436481099 0.31150744 0.2556729 0.23480666 0.1390845
#> [22,] 0.36891752 0.0269781852 0.35933632 0.2411360 0.13795694 0.1412164
#> [23,] 0.78326714 0.0341074318 0.70874767 0.5495387 0.45999600 0.3152802
#> [24,] 0.28967886 0.0319910811 0.20289127 0.1812264 0.10808045 0.1093417
#> [25,] 0.21056181 0.0134201885 0.14219914 0.1246671 0.13993079 0.1904267
#>              V44        V45        V46        V47        V48        V49
#>  [1,] 0.14143548 0.09352861 0.09653788 0.14094636 0.13665273 0.08162352
#>  [2,] 0.45071056 0.59278226 0.52340932 0.35614118 0.23016636 0.13522014
#>  [3,] 0.49720471 0.46362886 0.31612122 0.22043414 0.17801803 0.09942235
#>  [4,] 0.23164519 0.14738404 0.10961331 0.13044409 0.12286662 0.09560992
#>  [5,] 0.55571522 0.52379176 0.39000941 0.28798890 0.22627304 0.11968040
#>  [6,] 0.08955231 0.12653450 0.21210103 0.23014562 0.16613520 0.06917274
#>  [7,] 0.22914030 0.19410941 0.17657025 0.14549106 0.09000867 0.05474170
#>  [8,] 0.42875422 0.39902376 0.21880702 0.12881898 0.10722314 0.05456991
#>  [9,] 0.16414879 0.13054052 0.07151272 0.06377457 0.05849115 0.03825037
#> [10,] 0.02834109 0.02943791 0.06104752 0.06382492 0.09721342 0.04930270
#> [11,] 0.08595027 0.11930722 0.04662419 0.07746531 0.10791079 0.06630804
#> [12,] 0.27328819 0.33630813 0.29128899 0.19635036 0.11290616 0.07363533
#> [13,] 0.28740818 0.45304713 0.33756025 0.15526330 0.10061421 0.06141393
#> [14,] 0.21985761 0.13258963 0.10356023 0.06527712 0.04840986 0.03067789
#> [15,] 0.24924207 0.13231933 0.10064178 0.09902276 0.07193839 0.02825748
#> [16,] 0.15112619 0.08255177 0.07973616 0.03734039 0.03207262 0.02617637
#> [17,] 0.10520578 0.13374961 0.12385791 0.11124401 0.07206526 0.04221655
#> [18,] 0.17902173 0.18617825 0.16557121 0.14759756 0.13095724 0.08767964
#> [19,] 0.41041618 0.18666693 0.05760478 0.05990986 0.02992485 0.04282600
#> [20,] 0.24001274 0.16127428 0.12835360 0.08106325 0.07699902 0.02919041
#> [21,] 0.08040697 0.13058232 0.18467293 0.19653057 0.11289013 0.04509095
#> [22,] 0.16868510 0.18266316 0.10123905 0.05569118 0.05144246 0.02543423
#> [23,] 0.22653098 0.19419486 0.14028235 0.12215510 0.09869812 0.04230906
#> [24,] 0.08852547 0.06052868 0.05982600 0.03687287 0.03068390 0.01437453
#> [25,] 0.22454244 0.20250871 0.13597863 0.07539251 0.06980657 0.04170751
#>                V5         V50         V51         V52         V53         V54
#>  [1,] 0.063715879 0.023870278 0.018390598 0.014809108 0.011286869 0.016537960
#>  [2,] 0.082832472 0.045958737 0.026623454 0.020169180 0.015610939 0.012241749
#>  [3,] 0.090302191 0.024992115 0.012662286 0.009915057 0.011704572 0.006053768
#>  [4,] 0.123175702 0.029366723 0.029354142 0.008972273 0.014984245 0.013869031
#>  [5,] 0.064745845 0.036179002 0.023157086 0.018534805 0.008826944 0.008505748
#>  [6,] 0.099087598 0.006842547 0.022800406 0.008341464 0.005630454 0.006150744
#>  [7,] 0.094896697 0.017899634 0.021094346 0.011274133 0.010795865 0.009860016
#>  [8,] 0.093115360 0.019404937 0.017882375 0.012097726 0.013652020 0.011411977
#>  [9,] 0.137880200 0.016664232 0.017897565 0.016718406 0.012383830 0.015524734
#> [10,] 0.004496497 0.001854039 0.013028474 0.017264495 0.011309380 0.006587241
#> [11,] 0.048402325 0.002714067 0.012296173 0.014272896 0.003030968 0.012013269
#> [12,] 0.044284898 0.028086433 0.014786458 0.006592527 0.013602047 0.015886683
#> [13,] 0.072894818 0.033180926 0.018842361 0.016890881 0.013769392 0.022810324
#> [14,] 0.027257666 0.022167992 0.004279026 0.009614444 0.014249543 0.015176654
#> [15,] 0.080237676 0.020363387 0.028773093 0.011139439 0.011652965 0.012531122
#> [16,] 0.016606771 0.009962591 0.009638966 0.011221652 0.008975382 0.005298152
#> [17,] 0.038426987 0.026427455 0.015400552 0.006797623 0.008124539 0.009958129
#> [18,] 0.014921781 0.023232430 0.017607312 0.008907884 0.004051793 0.011894603
#> [19,] 0.068830729 0.022856647 0.005796608 0.007809195 0.013834414 0.018963343
#> [20,] 0.144093713 0.006812201 0.006584685 0.002148466 0.010829135 0.007869684
#> [21,] 0.133972105 0.016646575 0.008762534 0.006888030 0.006439554 0.008064102
#> [22,] 0.026722404 0.008484193 0.006441672 0.007311692 0.010791640 0.005465839
#> [23,] 0.013994508 0.024480276 0.022583748 0.007347682 0.006534903 0.007606378
#> [24,] 0.050282140 0.012346454 0.004651856 0.006318866 0.006162347 0.005110548
#> [25,] 0.013507285 0.014501793 0.006800835 0.006486980 0.004818147 0.010096250
#>               V55         V56         V57          V58          V59         V6
#>  [1,] 0.005584867 0.009130750 0.006998499 0.0097306569 0.0136372732 0.15233024
#>  [2,] 0.007979771 0.008063757 0.009578850 0.0098959740 0.0085615184 0.10807151
#>  [3,] 0.003668896 0.007343502 0.005429115 0.0049050468 0.0008919254 0.12822969
#>  [4,] 0.031444760 0.009101563 0.011783499 0.0225453446 0.0124750891 0.17813861
#>  [5,] 0.007548458 0.007756648 0.007223510 0.0041733394 0.0043176540 0.08349910
#>  [6,] 0.004889255 0.007743454 0.002429045 0.0051380095 0.0031841011 0.11204800
#>  [7,] 0.003291340 0.005398081 0.004744342 0.0025540720 0.0043119652 0.10479278
#>  [8,] 0.010001946 0.008497897 0.007704509 0.0070861681 0.0077936363 0.13072466
#>  [9,] 0.013361589 0.010669976 0.008148052 0.0099336618 0.0126696916 0.13401052
#> [10,] 0.008148915 0.010195256 0.004964375 0.0061647857 0.0061947835 0.04995940
#> [11,] 0.002714184 0.005349309 0.011043916 0.0073435655 0.0049822422 0.12081274
#> [12,] 0.015620093 0.011403852 0.008405347 0.0125430663 0.0096955042 0.05570207
#> [13,] 0.023284237 0.009897440 0.011442722 0.0129000878 0.0089345523 0.10341737
#> [14,] 0.010716443 0.011413144 0.008435054 0.0081926107 0.0068258143 0.04656338
#> [15,] 0.008563087 0.011125796 0.012520483 0.0037142160 0.0089557024 0.11644213
#> [16,] 0.004890930 0.002910136 0.006953381 0.0040986178 0.0033432210 0.04439260
#> [17,] 0.009975759 0.009177485 0.005961494 0.0033568550 0.0044867222 0.09124432
#> [18,] 0.008936651 0.003829481 0.012015661 0.0060596863 0.0087692562 0.09933625
#> [19,] 0.006405009 0.019589118 0.020205218 0.0103480267 0.0063053336 0.21440951
#> [20,] 0.012962143 0.007539183 0.007964013 0.0074118679 0.0029734231 0.19378445
#> [21,] 0.011718700 0.008729943 0.010217418 0.0073384703 0.0042295137 0.14061813
#> [22,] 0.004971653 0.007318289 0.003354148 0.0008075113 0.0053966688 0.02518576
#> [23,] 0.010521581 0.008224418 0.003425560 0.0049173595 0.0049087954 0.06100522
#> [24,] 0.004576502 0.003648951 0.005724563 0.0051145784 0.0039678292 0.07760534
#> [25,] 0.006583912 0.004253024 0.006106608 0.0043445360 0.0071844149 0.07808522
#>                V60         V7          V8         V9
#>  [1,] 0.0115204846 0.17281611 0.173666136 0.21366026
#>  [2,] 0.0058584010 0.10663646 0.135709388 0.23624045
#>  [3,] 0.0031531860 0.16150313 0.148565188 0.29801118
#>  [4,] 0.0042237529 0.26580233 0.298569697 0.29301378
#>  [5,] 0.0026002038 0.11474645 0.133270458 0.26158593
#>  [6,] 0.0006819212 0.07324962 0.098253505 0.16396320
#>  [7,] 0.0047361923 0.10229901 0.140910734 0.20680506
#>  [8,] 0.0052135731 0.15333137 0.169343426 0.30046305
#>  [9,] 0.0094594803 0.12385125 0.118611429 0.13747567
#> [10,] 0.0078188867 0.04523464 0.009652656 0.12928613
#> [11,] 0.0094769090 0.06602964 0.050464755 0.14835726
#> [12,] 0.0063044386 0.06225254 0.128711401 0.25907570
#> [13,] 0.0114233560 0.25008260 0.396932224 0.53877980
#> [14,] 0.0105275662 0.08446436 0.103950456 0.07454873
#> [15,] 0.0044935588 0.13489950 0.142194813 0.14164620
#> [16,] 0.0041905139 0.02895518 0.055693377 0.08159605
#> [17,] 0.0066364740 0.11759000 0.080065026 0.12536583
#> [18,] 0.0056993427 0.13259047 0.242392168 0.30137236
#> [19,] 0.0053146317 0.15077509 0.111305366 0.08440558
#> [20,] 0.0146073432 0.14579566 0.149304847 0.04503358
#> [21,] 0.0055009774 0.11538933 0.121825163 0.21385317
#> [22,] 0.0037933377 0.04310425 0.025701633 0.03669843
#> [23,] 0.0041171191 0.06489859 0.091545212 0.05345654
#> [24,] 0.0042327638 0.08355722 0.091662154 0.12846613
#> [25,] 0.0042075687 0.07568695 0.046875327 0.06147248
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3623188