Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.084283607 0.13286407 0.19634408 0.29868655 0.32272861 0.23826760
#>  [2,] 0.015593871 0.32860741 0.29122823 0.29159873 0.23508729 0.16376940
#>  [3,] 0.040292085 0.41081135 0.50301413 0.52988253 0.52730717 0.39661483
#>  [4,] 0.024187547 0.25780944 0.31021823 0.34212131 0.29845517 0.20952685
#>  [5,] 0.043000000 0.71060000 0.73420000 0.50330000 0.30000000 0.19510000
#>  [6,] 0.009180355 0.19545650 0.20522296 0.14900848 0.05726285 0.14815733
#>  [7,] 0.036021796 0.03959278 0.17628471 0.30912548 0.34585374 0.42900858
#>  [8,] 0.010998325 0.41305634 0.47335381 0.36377172 0.31117546 0.22206388
#>  [9,] 0.033306663 0.26542318 0.30884052 0.30934552 0.26234245 0.18886497
#> [10,] 0.056071692 0.28923877 0.32310754 0.33204326 0.47595909 0.64094108
#> [11,] 0.042909003 0.21599978 0.28193230 0.29284646 0.36946589 0.34490656
#> [12,] 0.051385421 0.36483460 0.34271411 0.25229818 0.27038678 0.29716831
#> [13,] 0.025095661 0.21522816 0.25844938 0.29538380 0.32649514 0.39449116
#> [14,] 0.014581845 0.12336608 0.12431850 0.10710152 0.11828237 0.13662376
#> [15,] 0.016930536 0.11536981 0.13501992 0.18499643 0.24860834 0.22993199
#> [16,] 0.003499608 0.12606453 0.12949020 0.15331771 0.21065671 0.34721712
#> [17,] 0.009300000 0.13600000 0.21320000 0.37380000 0.37380000 0.26730000
#> [18,] 0.020831065 0.10958520 0.13020835 0.13810493 0.21046339 0.34461597
#> [19,] 0.030930758 0.31151081 0.27222461 0.15963372 0.24332445 0.46839788
#> [20,] 0.001221567 0.02813443 0.04544852 0.08196187 0.11388304 0.10815576
#> [21,] 0.025284063 0.06482847 0.02545809 0.04476794 0.08452624 0.08978406
#> [22,] 0.021100000 0.05680000 0.08690000 0.19350000 0.14780000 0.18710000
#> [23,] 0.005059900 0.08106151 0.06583786 0.18456601 0.04164581 0.21513846
#> [24,] 0.029920024 0.05880310 0.07713369 0.15019460 0.17577135 0.19377340
#> [25,] 0.010932937 0.10427638 0.10060490 0.10011162 0.11408230 0.12788214
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.23440257 0.28444097 0.22936897 0.2853972 0.3925292 0.145845243
#>  [2,] 0.16945035 0.21339780 0.20355115 0.2158872 0.1693878 0.019507595
#>  [3,] 0.28745177 0.31780627 0.32376479 0.3132726 0.5059594 0.056751395
#>  [4,] 0.18338852 0.23762287 0.24250849 0.3591115 0.5626915 0.032448625
#>  [5,] 0.27670000 0.37370000 0.25070000 0.2507000 0.3292000 0.090200000
#>  [6,] 0.26192652 0.55132366 0.70885368 0.7955000 0.8028547 0.012410221
#>  [7,] 0.53910026 0.61224600 0.75408542 0.8105576 0.8593699 0.035806234
#>  [8,] 0.09148336 0.05923461 0.21564465 0.4148012 0.6096310 0.007486632
#>  [9,] 0.18214603 0.18397394 0.20654075 0.2263295 0.2865052 0.037536270
#> [10,] 0.68538325 0.69358216 0.70159974 0.7007770 0.7318488 0.065326755
#> [11,] 0.28147398 0.30987504 0.34792934 0.2242655 0.3641417 0.071939946
#> [12,] 0.28770872 0.29488234 0.23646426 0.2490483 0.3505055 0.074173144
#> [13,] 0.45674715 0.52448478 0.60193494 0.6743346 0.7320824 0.025930124
#> [14,] 0.12099478 0.13713684 0.19370664 0.2538188 0.3127374 0.015500757
#> [15,] 0.24315806 0.30585272 0.34434039 0.3587456 0.4209773 0.022856771
#> [16,] 0.28597687 0.25109918 0.40357601 0.6860359 0.7737726 0.005383661
#> [17,] 0.23330000 0.53670000 0.73120000 0.7659000 0.6271000 0.026900000
#> [18,] 0.49259771 0.64778930 0.77114504 0.8130233 0.8300663 0.021443580
#> [19,] 0.60361085 0.70397087 0.77155861 0.6741425 0.4473390 0.061177297
#> [20,] 0.05043356 0.09576569 0.13818742 0.2320224 0.1735618 0.004803055
#> [21,] 0.13516137 0.06781189 0.10917459 0.2167332 0.3291604 0.061967945
#> [22,] 0.19940000 0.32830000 0.68610000 0.5814000 0.2500000 0.031900000
#> [23,] 0.15859887 0.21643020 0.07727311 0.2361805 0.4248640 0.014949414
#> [24,] 0.15779097 0.13099152 0.15475246 0.1007649 0.1187229 0.048781918
#> [25,] 0.17496213 0.18038135 0.14800849 0.1551802 0.1131888 0.014143738
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.2179905 0.2534631 0.5688264 0.8696922 0.9137290 0.7500796 0.5805211
#>  [2,] 0.1222729 0.1252548 0.2553896 0.4415789 0.5979665 0.7241392 0.8441644
#>  [3,] 0.7367566 0.8250692 0.8224715 0.7288622 0.7487925 0.7756570 0.8886797
#>  [4,] 0.6878639 0.7885147 0.7744742 0.8078257 0.9156392 0.9076578 0.9411385
#>  [5,] 0.4871000 0.6527000 0.8454000 0.9739000 1.0000000 0.6665000 0.5323000
#>  [6,] 0.6958304 0.7250840 0.8801759 0.8982570 0.4817199 0.1461543 0.2876681
#>  [7,] 0.9199422 0.8754389 0.8081403 0.8167355 0.8728801 0.7710610 0.5600338
#>  [8,] 0.7011754 0.7758049 0.8475178 0.8574387 0.8776515 0.8683643 0.8818525
#>  [9,] 0.3184736 0.4677929 0.6296297 0.6945322 0.7486163 0.8118949 0.8522814
#> [10,] 0.8060339 0.8819696 0.8213635 0.7515878 0.5727567 0.4072086 0.5242220
#> [11,] 0.4592699 0.5809697 0.6303242 0.6187431 0.6400277 0.5843641 0.5787048
#> [12,] 0.7238307 0.7943947 0.5098940 0.3819658 0.3771839 0.3921943 0.6282289
#> [13,] 0.7835186 0.8382964 0.9015603 0.9368807 0.9696204 0.9401601 0.9217756
#> [14,] 0.4006913 0.4864415 0.6214390 0.7410665 0.7716946 0.7698604 0.8093699
#> [15,] 0.4754390 0.5255496 0.5466586 0.5699966 0.6357288 0.7713285 0.8299565
#> [16,] 0.9042812 0.9790531 0.9444132 0.8957184 0.8110209 0.7495583 0.5993300
#> [17,] 0.4395000 0.4330000 0.4326000 0.5544000 0.7360000 0.8589000 0.8989000
#> [18,] 0.8149461 0.8817798 0.9014467 0.8751674 0.8734099 0.8791186 0.8764574
#> [19,] 0.3971699 0.5301273 0.6893105 0.7862331 0.8227955 0.6979885 0.6108148
#> [20,] 0.1362862 0.1737156 0.3248488 0.3765525 0.5061687 0.5317761 0.6415701
#> [21,] 0.3017278 0.1751768 0.2190566 0.3489913 0.4358157 0.4761812 0.2594379
#> [22,] 0.1734000 0.3363000 0.5588000 0.6592000 0.7012000 0.8099000 0.8901000
#> [23,] 0.3279612 0.2643539 0.3924701 0.5558183 0.5234581 0.3675500 0.5598876
#> [24,] 0.1224828 0.1448510 0.2018035 0.1603031 0.2040615 0.3334082 0.5147449
#> [25,] 0.1351785 0.2177740 0.2468490 0.3781945 0.5998002 0.8403561 0.9812801
#>              V27        V28       V29           V3       V30       V31
#>  [1,] 0.47704970 0.54757969 0.7186475  0.193286852 0.5044006 0.4834477
#>  [2,] 0.91531720 0.93226254 0.9018296  0.018753367 0.9287182 0.9018712
#>  [3,] 0.98043668 0.91257474 0.7568891  0.045827616 0.5211571 0.4414299
#>  [4,] 0.97557932 0.97682554 0.8582427  0.037481847 0.6369128 0.4256854
#>  [5,] 0.40240000 0.34440000 0.4239000  0.083300000 0.4182000 0.4393000
#>  [6,] 0.63662742 0.86260166 0.9628358 -0.014306258 0.8213307 0.4792301
#>  [7,] 0.41866342 0.33096528 0.3689118  0.062095221 0.3587871 0.3295129
#>  [8,] 0.95140480 0.92753809 0.7356566  0.033478523 0.5121241 0.2778747
#>  [9,] 0.92645524 0.92764866 0.8395649  0.042372818 0.7774482 0.7454963
#> [10,] 0.52998876 0.46509762 0.3871946  0.079097928 0.3431995 0.2152107
#> [11,] 0.61087167 0.77348304 0.8438468  0.072967031 0.8774181 0.9252178
#> [12,] 0.66895545 0.71600994 0.6680092  0.064898497 0.6701237 0.4392419
#> [13,] 0.90637655 0.84086287 0.7342093  0.037358796 0.5921519 0.4156324
#> [14,] 0.80078434 0.82347815 0.8696654  0.021265689 0.8101070 0.7100352
#> [15,] 0.84477209 0.78692515 0.6548615  0.025709159 0.5179065 0.4368963
#> [16,] 0.34879763 0.08698241 0.4154184  0.005647955 0.2188356 0.1913183
#> [17,] 0.94200000 0.94010000 0.9379000  0.021700000 0.8575000 0.7284000
#> [18,] 0.75637502 0.56968202 0.4646525  0.032843803 0.4831085 0.4290816
#> [19,] 0.67308587 0.60874077 0.4186757  0.088246238 0.4245260 0.4659063
#> [20,] 0.72118409 0.82152072 0.9227765  0.001961112 0.8492556 0.6633269
#> [21,] 0.08391827 0.28719813 0.5654782  0.036769445 0.8430317 0.8341923
#> [22,] 0.87450000 0.78870000 0.8725000  0.041500000 0.9376000 0.8920000
#> [23,] 0.52880687 0.27328701 0.2806357  0.055808278 0.4339089 0.6874326
#> [24,] 0.59610941 0.61436254 0.5946313  0.030681951 0.6159620 0.7457536
#> [25,] 0.83608669 0.61753113 0.5071661  0.012050200 0.3836246 0.2986136
#>              V32       V33       V34        V35        V36        V37
#>  [1,] 0.38646603 0.3526454 0.1815878 0.07366389 0.42877926 0.51556878
#>  [2,] 0.80708221 0.7000103 0.6450990 0.64730604 0.69963292 0.67199246
#>  [3,] 0.49257262 0.4258701 0.2872604 0.08483533 0.27072216 0.26267372
#>  [4,] 0.28473261 0.2136379 0.1630196 0.13542059 0.04085381 0.07216172
#>  [5,] 0.11620000 0.4336000 0.6553000 0.61720000 0.43730000 0.41180000
#>  [6,] 0.45342486 0.4200367 0.3809547 0.27946356 0.19280954 0.19962922
#>  [7,] 0.35202495 0.4342626 0.4931150 0.49166693 0.36561032 0.27690167
#>  [8,] 0.33491137 0.4927603 0.5942374 0.64752970 0.70166316 0.66284093
#>  [9,] 0.58750273 0.4058906 0.3027161 0.18349537 0.12574946 0.19817643
#> [10,] 0.17853670 0.1766934 0.1191510 0.13415044 0.14378010 0.17503651
#> [11,] 0.83166397 0.6192758 0.5750847 0.55797816 0.31243059 0.17822337
#> [12,] 0.45510510 0.5167294 0.5575046 0.53282969 0.59987320 0.57036932
#> [13,] 0.30331274 0.2212639 0.1222257 0.15781173 0.15005185 0.19036701
#> [14,] 0.64152220 0.6451880 0.6677920 0.65682781 0.57381339 0.44307240
#> [15,] 0.37442582 0.2910462 0.2259197 0.26545696 0.33661844 0.34864155
#> [16,] 0.02544941 0.1010639 0.2682380 0.34115414 0.42310138 0.36230510
#> [17,] 0.67000000 0.7547000 0.8773000 0.99190000 0.99220000 0.94190000
#> [18,] 0.33220737 0.2991903 0.3086035 0.20729396 0.12298117 0.15013991
#> [19,] 0.34687378 0.2353433 0.1461065 0.39965788 0.61174290 0.61974908
#> [20,] 0.57205410 0.7225798 0.5935831 0.45332620 0.45738275 0.42432616
#> [21,] 0.74352861 0.8173492 0.9566584 1.00643862 0.89793533 0.54067215
#> [22,] 0.75080000 0.6832000 0.7610000 0.90170000 1.00000000 0.91230000
#> [23,] 0.70072505 0.5637096 0.3571886 0.41518714 0.34362473 0.53502000
#> [24,] 0.74938983 0.7514840 0.7636931 0.89732700 0.98386370 0.87658682
#> [25,] 0.15893145 0.1751948 0.2084685 0.25795203 0.19458420 0.06920077
#>              V38        V39          V4       V40        V41        V42
#>  [1,] 0.26538286 0.23739278 0.202789388 0.3008014 0.29841932 0.41263388
#>  [2,] 0.55565348 0.46489746 0.051009346 0.4811684 0.46727304 0.42520760
#>  [3,] 0.38621070 0.49923877 0.032052768 0.3927202 0.22544148 0.41782215
#>  [4,] 0.08960980 0.09620280 0.047968296 0.1125674 0.06813366 0.10087034
#>  [5,] 0.36410000 0.45720000 0.081300000 0.4367000 0.29640000 0.43120000
#>  [6,] 0.44600823 0.45848978 0.004938914 0.3433263 0.08388452 0.14900328
#>  [7,] 0.35070845 0.41222396 0.073291001 0.4018968 0.31403012 0.28271262
#>  [8,] 0.53910726 0.42859005 0.051770387 0.3299765 0.25947054 0.18448683
#>  [9,] 0.18087091 0.18458963 0.055131772 0.2146849 0.25692367 0.19000623
#> [10,] 0.18351901 0.21690921 0.098903440 0.1651176 0.14758381 0.22978507
#> [11,] 0.27616705 0.31598703 0.066706483 0.3485749 0.63464371 0.64832725
#> [12,] 0.62499987 0.65418379 0.071902231 0.4872046 0.39828806 0.40558825
#> [13,] 0.24396909 0.21895074 0.052298751 0.2378353 0.27387006 0.23970406
#> [14,] 0.31820001 0.22611477 0.028573232 0.2378418 0.23211383 0.19102458
#> [15,] 0.29394233 0.29549355 0.030136162 0.2433562 0.22292782 0.22470004
#> [16,] 0.12653857 0.07770284 0.022879314 0.1089947 0.11134139 0.09194900
#> [17,] 0.83880000 0.66050000 0.033900000 0.4816000 0.29170000 0.17690000
#> [18,] 0.14173893 0.12560378 0.046918870 0.1736515 0.14758710 0.09797922
#> [19,] 0.38481471 0.24236122 0.080315030 0.3521890 0.37575133 0.32188043
#> [20,] 0.46123521 0.47835693 0.023304641 0.3787220 0.15420333 0.05034175
#> [21,] 0.27568216 0.20339228 0.016433082 0.2208429 0.33418051 0.39575981
#> [22,] 0.73880000 0.59150000 0.028600000 0.4057000 0.30190000 0.23310000
#> [23,] 0.84488581 0.97371245 0.019999407 0.9365707 0.66956645 0.57614786
#> [24,] 0.72243381 0.58940274 0.027071426 0.4869113 0.39545865 0.34591480
#> [25,] 0.06638803 0.14644220 0.017619185 0.2149800 0.28018609 0.27156576
#>              V43        V44        V45         V46          V47        V48
#>  [1,] 0.28849588 0.10268290 0.15031041  0.21555217  0.298735118 0.26673314
#>  [2,] 0.43308104 0.50178062 0.47426319  0.32949393  0.229988405 0.19084279
#>  [3,] 0.43594499 0.46527182 0.44297056  0.18596521  0.104855771 0.13173339
#>  [4,] 0.19769372 0.22333642 0.19966179  0.20600620  0.165053946 0.09978327
#>  [5,] 0.41550000 0.18240000 0.14870000  0.01380000  0.116400000 0.20520000
#>  [6,] 0.42561698 0.35155678 0.09031060  0.17774521  0.149829436 0.07626666
#>  [7,] 0.26252652 0.21575784 0.16965763  0.09987400  0.081849173 0.05373891
#>  [8,] 0.25574283 0.29711580 0.24775177  0.18178383  0.090359292 0.05914627
#>  [9,] 0.16999034 0.13071935 0.12185510  0.09198521  0.102063871 0.11003014
#> [10,] 0.24772737 0.18099067 0.12282917  0.10090997  0.066865553 0.05871230
#> [11,] 0.43230709 0.40470978 0.58455681  0.55159279  0.328505616 0.18112065
#> [12,] 0.37123660 0.32599537 0.45479959  0.33989308  0.187812084 0.17853363
#> [13,] 0.18053482 0.17075196 0.14319225  0.09803882  0.091969158 0.07154998
#> [14,] 0.20083119 0.16827653 0.11810037  0.08339956  0.051777884 0.03857589
#> [15,] 0.19109152 0.15942658 0.10373884  0.07888156  0.059539013 0.04913025
#> [16,] 0.14140725 0.08103431 0.08722152  0.10328712  0.143653558 0.07962043
#> [17,] 0.11360000 0.07010000 0.15780000  0.19380000  0.110600000 0.06930000
#> [18,] 0.06317611 0.10803450 0.10577525  0.05830259  0.082083204 0.06423204
#> [19,] 0.22474517 0.17831088 0.17551457  0.12138729  0.127093993 0.11382821
#> [20,] 0.07525983 0.07719458 0.06506778 -0.01288897 -0.002210611 0.02909796
#> [21,] 0.15936504 0.22061631 0.14118030  0.13018377  0.160091655 0.10957460
#> [22,] 0.29310000 0.22980000 0.23910000  0.19100000  0.109600000 0.03000000
#> [23,] 0.42571268 0.41397238 0.45625336  0.37792903  0.276603279 0.16048481
#> [24,] 0.32250503 0.27690366 0.22174484  0.15186034  0.108214703 0.09226817
#> [25,] 0.20398575 0.11747401 0.09737291  0.10058688  0.049437120 0.03640794
#>               V49         V5         V50         V51         V52         V53
#>  [1,] 0.131854402 0.19362701 0.020754115 0.032160757 0.042824303 0.022615684
#>  [2,] 0.103881520 0.07708653 0.027798686 0.016729544 0.013489899 0.011161578
#>  [3,] 0.083090326 0.05662790 0.017800782 0.029385129 0.012715770 0.004466559
#>  [4,] 0.049120881 0.05523580 0.013749687 0.017478331 0.013112071 0.005599002
#>  [5,] 0.106900000 0.01650000 0.019900000 0.020800000 0.017600000 0.019700000
#>  [6,] 0.065788920 0.01938042 0.026577028 0.012080537 0.017569216 0.008810796
#>  [7,] 0.028787372 0.10887250 0.003915283 0.002390131 0.006360612 0.014078139
#>  [8,] 0.030693165 0.04886655 0.015723843 0.005534214 0.012175453 0.005829416
#>  [9,] 0.063588644 0.05827680 0.014948449 0.014203587 0.014362667 0.006090645
#> [10,] 0.052100460 0.11679737 0.021986398 0.022170645 0.014529087 0.014560782
#> [11,] 0.098310813 0.08726939 0.043925640 0.033948689 0.028071456 0.018815155
#> [12,] 0.107937744 0.07116022 0.038866568 0.028348921 0.019717670 0.011993482
#> [13,] 0.048563155 0.10137792 0.022292150 0.016423202 0.017540489 0.015806920
#> [14,] 0.016531879 0.04851248 0.012168817 0.008027561 0.007121171 0.008105510
#> [15,] 0.023140818 0.02892653 0.010786382 0.007987694 0.008170351 0.009361510
#> [16,] 0.022222769 0.03314476 0.030137732 0.017490629 0.018288025 0.016001649
#> [17,] 0.017600000 0.03050000 0.020500000 0.030900000 0.021200000 0.009100000
#> [18,] 0.028574819 0.08527011 0.012188586 0.010765236 0.008742842 0.007417261
#> [19,] 0.076298396 0.08002225 0.021053347 0.017706647 0.012679773 0.008368787
#> [20,] 0.005191054 0.03113833 0.004536170 0.004292204 0.008489754 0.018152267
#> [21,] 0.054239829 0.04874339 0.011547421 0.016728210 0.003200093 0.005139081
#> [22,] 0.017100000 0.01210000 0.038300000 0.005300000 0.009000000 0.004200000
#> [23,] 0.070375639 0.04217330 0.026035454 0.023509572 0.011775558 0.005684347
#> [24,] 0.051746705 0.04174077 0.010612962 0.009288032 0.011070383 0.019449239
#> [25,] 0.030031157 0.03529418 0.009854210 0.006045494 0.009193376 0.003915214
#>               V54         V55         V56          V57          V58         V59
#>  [1,] 0.026386649 0.014849417 0.013256059  0.016335677 0.0181125401 0.020142741
#>  [2,] 0.006428135 0.005131993 0.006516445  0.005756724 0.0048401738 0.002233457
#>  [3,] 0.008599901 0.009408511 0.005905672  0.006162603 0.0052016044 0.003851598
#>  [4,] 0.005789276 0.004248263 0.003812252  0.004074983 0.0031045209 0.002033272
#>  [5,] 0.021000000 0.014100000 0.004900000  0.002700000 0.0162000000 0.005900000
#>  [6,] 0.002745009 0.003243490 0.004479970 -0.002475422 0.0039641665 0.002095980
#>  [7,] 0.010673928 0.009715336 0.017267009  0.008036221 0.0079836414 0.014566111
#>  [8,] 0.011796944 0.007062524 0.006515613  0.003296170 0.0088616434 0.010438430
#>  [9,] 0.008907897 0.005624502 0.006742635  0.007200908 0.0070606244 0.005245118
#> [10,] 0.013623942 0.008794299 0.006843943  0.007968273 0.0121468981 0.009864290
#> [11,] 0.016890660 0.014785889 0.011069537  0.007434339 0.0120613271 0.013286772
#> [12,] 0.013855640 0.014669777 0.010581844  0.008183761 0.0078943748 0.007724677
#> [13,] 0.013312869 0.007098522 0.011710697  0.008891689 0.0061767381 0.006060101
#> [14,] 0.007265059 0.005011138 0.005653679  0.006835544 0.0050055819 0.006033578
#> [15,] 0.010981479 0.009186227 0.007783872  0.006881266 0.0070834565 0.005940353
#> [16,] 0.008327467 0.011643264 0.001259976  0.006715109 0.0049369669 0.008057554
#> [17,] 0.005600000 0.008600000 0.009200000  0.007000000 0.0116000000 0.006000000
#> [18,] 0.007398541 0.006569421 0.009886864  0.006615663 0.0026845695 0.004738403
#> [19,] 0.012679487 0.008615795 0.005959345  0.009044453 0.0080974666 0.010089299
#> [20,] 0.008969237 0.011339982 0.006782190  0.002079539 0.0008665759 0.005865073
#> [21,] 0.008290628 0.007387628 0.007335989  0.010101687 0.0160890348 0.009626523
#> [22,] 0.015300000 0.010600000 0.002000000  0.010500000 0.0049000000 0.007000000
#> [23,] 0.014432966 0.008786906 0.009480822  0.003435899 0.0067130385 0.008023739
#> [24,] 0.008142075 0.014924295 0.008201497  0.006523959 0.0096488239 0.012054350
#> [25,] 0.004080718 0.005825305 0.002876974  0.002896234 0.0033509431 0.001572358
#>               V6          V60         V7           V8          V9
#>  [1,] 0.25818771 0.0090435526 0.24201049  0.139359387 0.074712744
#>  [2,] 0.11409967 0.0029306972 0.14180422  0.141866138 0.286412211
#>  [3,] 0.12118365 0.0050416579 0.12930623  0.187096478 0.291593767
#>  [4,] 0.09245888 0.0037052590 0.13125489  0.144348807 0.189258413
#>  [5,] 0.02770000 0.0021000000 0.05690000  0.205700000 0.388700000
#>  [6,] 0.03801131 0.0053360832 0.05673924 -0.005364006 0.092266052
#>  [7,] 0.12798251 0.0077883306 0.13688013  0.102796898 0.063817396
#>  [8,] 0.05843941 0.0063214737 0.14842047  0.212398275 0.356359092
#>  [9,] 0.10301789 0.0079193520 0.10187303  0.111567477 0.188226521
#> [10,] 0.14074993 0.0075790458 0.16675262  0.208905139 0.247831208
#> [11,] 0.16262711 0.0103931447 0.13730036  0.152134802 0.255097334
#> [12,] 0.09068894 0.0074319298 0.11435329  0.200306549 0.321909956
#> [13,] 0.13608875 0.0042845502 0.15128009  0.155877495 0.198509640
#> [14,] 0.07144046 0.0050497344 0.07649688  0.079623446 0.102419959
#> [15,] 0.05325146 0.0046172422 0.08210620  0.083036559 0.093174549
#> [16,] 0.04620828 0.0036598522 0.02797742  0.079002870 0.083315549
#> [17,] 0.11720000 0.0110000000 0.14500000  0.063800000 0.074000000
#> [18,] 0.10683536 0.0048457987 0.11899265  0.089370508 0.105632418
#> [19,] 0.11062294 0.0079670461 0.14014815  0.194621728 0.251773043
#> [20,] 0.03270439 0.0050790493 0.03413663  0.026013750 0.050003439
#> [21,] 0.07948660 0.0007485469 0.09262708  0.020506548 0.056683393
#> [22,] 0.04380000 0.0080000000 0.12990000  0.139000000 0.069500000
#> [23,] 0.04874527 0.0131790517 0.10975966  0.134131092 0.036027694
#> [24,] 0.07955591 0.0055199157 0.08672762  0.052220681 0.002056702
#> [25,] 0.05830175 0.0026590238 0.07343993  0.070651064 0.109513487
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913