Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11         V12        V13        V14
#>  [1,] 0.025090121 0.34386404 0.32657573  0.35503979 0.38355544 0.32662880
#>  [2,] 0.006406394 0.22757023 0.07364902 -0.01512730 0.21292976 0.34357595
#>  [3,] 0.053423584 0.14144726 0.21845867  0.24269993 0.32210113 0.33367303
#>  [4,] 0.070700000 0.12470000 0.23400000  0.17640000 0.22840000 0.31150000
#>  [5,] 0.042606379 0.15875699 0.26768738  0.38867900 0.41634424 0.45782507
#>  [6,] 0.010834885 0.35381808 0.31770698  0.31224317 0.23473156 0.14020309
#>  [7,] 0.027985533 0.33438476 0.41447819  0.44536473 0.41783930 0.32571370
#>  [8,] 0.027450592 0.22830704 0.28134741  0.29642121 0.25185635 0.17340756
#>  [9,] 0.017233707 0.60149946 0.63272695  0.45432155 0.37245469 0.28969294
#> [10,] 0.016612051 0.11031951 0.17228572  0.19615848 0.18594846 0.12584785
#> [11,] 0.020829463 0.19638341 0.18415831  0.15348845 0.15082447 0.16468646
#> [12,] 0.031050740 0.21519149 0.25643440  0.30317343 0.35628573 0.39992117
#> [13,] 0.035695885 0.40170021 0.38507969  0.40101095 0.51220394 0.67014991
#> [14,] 0.026701908 0.07600942 0.11531612  0.10693117 0.11293370 0.21220292
#> [15,] 0.011900000 0.08640000 0.21430000  0.37200000 0.26650000 0.21130000
#> [16,] 0.004383233 0.14805245 0.07751474  0.09615894 0.15414474 0.21310853
#> [17,] 0.026091941 0.06964197 0.03816071  0.07967346 0.11363359 0.11710011
#> [18,] 0.017705548 0.11398230 0.09721836  0.10254916 0.11842597 0.13293652
#> [19,] 0.034280959 0.24311843 0.20329272  0.23693363 0.25212730 0.11907471
#> [20,] 0.009289304 0.09841319 0.11022102  0.13757895 0.21439875 0.17892386
#> [21,] 0.014005857 0.09690495 0.10874513  0.14295016 0.13092988 0.13037757
#> [22,] 0.025451980 0.06305152 0.14430731  0.32251497 0.29031950 0.26463689
#> [23,] 0.023884670 0.15196314 0.22456439  0.16841886 0.23627543 0.36754561
#> [24,] 0.014408101 0.05991032 0.10675473  0.07027111 0.12785145 0.17344874
#> [25,] 0.018999040 0.10654134 0.07913615  0.05215348 0.07510436 0.08649796
#>              V15        V16       V17       V18       V19           V2
#>  [1,] 0.28971358 0.28201666 0.2327752 0.1732010 0.3207012  0.024447814
#>  [2,] 0.47039115 0.61583878 0.6419682 0.7009270 0.8246532  0.003656739
#>  [3,] 0.29933036 0.27756822 0.3289753 0.1656527 0.1939792  0.069501838
#>  [4,] 0.47250000 0.55430000 0.5386000 0.3746000 0.4583000  0.125200000
#>  [5,] 0.53820276 0.61892931 0.7606129 0.8532933 0.9322525  0.062870039
#>  [6,] 0.18071505 0.23876566 0.2083184 0.2128265 0.1627971  0.009888940
#>  [7,] 0.25137938 0.27978937 0.3053198 0.3437526 0.5085011  0.038085139
#>  [8,] 0.15841429 0.18367442 0.2110623 0.2392166 0.2883036  0.032485419
#>  [9,] 0.16362011 0.03764927 0.1724059 0.4358419 0.6954669  0.012754669
#> [10,] 0.06675012 0.14139304 0.2222935 0.3322863 0.4215336  0.033874916
#> [11,] 0.29067028 0.41509615 0.5161772 0.5467682 0.6252527  0.028039208
#> [12,] 0.44022379 0.43256400 0.5498726 0.6201342 0.6319907  0.031079673
#> [13,] 0.69500411 0.72946902 0.7567773 0.8149564 0.8737210  0.083595086
#> [14,] 0.26653823 0.41643673 0.4980260 0.6409680 0.7979267  0.008537308
#> [15,] 0.11030000 0.11360000 0.1934000 0.4142000 0.3279000  0.058200000
#> [16,] 0.26305328 0.34178590 0.3188186 0.3027987 0.3293543 -0.001695832
#> [17,] 0.14735628 0.08564779 0.1203123 0.1960774 0.2377775  0.053316265
#> [18,] 0.14859035 0.16693664 0.1379301 0.1472795 0.1221606  0.015937144
#> [19,] 0.32188726 0.30860014 0.2495272 0.2421877 0.2433016  0.011286765
#> [20,] 0.16661368 0.24450413 0.3386749 0.3912424 0.4406020  0.010901021
#> [21,] 0.14502307 0.23655139 0.3554573 0.3907215 0.3811369  0.019535890
#> [22,] 0.05021658 0.10545189 0.2387858 0.5057649 0.4892408  0.016288311
#> [23,] 0.51595759 0.61305275 0.5280974 0.5019857 0.5776415  0.025168179
#> [24,] 0.15774242 0.18075206 0.1675518 0.1519055 0.1207723  0.006434043
#> [25,] 0.01489199 0.10666706 0.1424078 0.3591429 0.3397695  0.016724863
#>             V20       V21       V22       V23       V24        V25       V26
#>  [1,] 0.4733611 0.5283894 0.4516003 0.4706647 0.5086934 0.58512685 0.7054468
#>  [2,] 0.9100750 1.0246862 1.0291146 0.9434558 0.6739467 0.45894770 0.4797650
#>  [3,] 0.3352627 0.4772562 0.5835229 0.6173749 0.6218652 0.55870704 0.5169654
#>  [4,] 0.5961000 0.7464000 0.7644000 0.5711000 0.6257000 0.66950000 0.7131000
#>  [5,] 0.9284483 0.9392213 0.8009131 0.6613686 0.6620799 0.55473320 0.3768721
#>  [6,] 0.1207286 0.1393254 0.2636523 0.4598655 0.6190365 0.73556129 0.8475188
#>  [7,] 0.6668466 0.7948745 0.8330569 0.8213794 0.8666625 0.88303020 0.9195244
#>  [8,] 0.3332550 0.4438339 0.5611189 0.6368586 0.7189098 0.79723373 0.8635345
#>  [9,] 0.8099901 0.8555235 0.8737487 0.8600433 0.9242519 0.88334085 0.8208375
#> [10,] 0.5524329 0.6269418 0.6937307 0.7024697 0.6254687 0.68520622 0.8247369
#> [11,] 0.7991694 0.9395174 1.0018888 0.8888773 0.7084600 0.52732383 0.3500753
#> [12,] 0.6789866 0.7644693 0.8349340 0.9080083 0.9688745 0.99779559 0.9731641
#> [13,] 0.9163092 0.8600382 0.8857377 1.0229879 0.9930186 0.82862540 0.8023711
#> [14,] 0.8900790 0.9177658 0.9767981 1.0237043 0.9941676 0.95733251 0.8738897
#> [15,] 0.6222000 0.7468000 0.7676000 0.7867000 0.8253000 1.00000000 0.9481000
#> [16,] 0.4259491 0.5090235 0.5695749 0.5998704 0.5478269 0.65953766 0.8591293
#> [17,] 0.2280883 0.1852281 0.2255173 0.2454332 0.3349238 0.39992656 0.3294578
#> [18,] 0.1329404 0.2368832 0.2921871 0.4370891 0.6450574 0.91463084 1.0232159
#> [19,] 0.2816582 0.1941850 0.5711799 0.5652502 0.3319511 0.27769446 0.5360407
#> [20,] 0.4835419 0.5727940 0.6315368 0.6107662 0.6808067 0.82936759 0.8512815
#> [21,] 0.4275591 0.4861870 0.6007658 0.6908844 0.7621539 0.72897178 0.6902027
#> [22,] 0.6277659 0.6772239 0.7458932 0.8408417 0.8698534 0.91257829 0.7159341
#> [23,] 0.5549762 0.5239090 0.6081052 0.7901559 0.9037429 0.84019025 0.8356775
#> [24,] 0.1265390 0.1860615 0.3626913 0.5201087 0.6349614 0.70261107 0.8031842
#> [25,] 0.3212277 0.2310124 0.1989048 0.1819301 0.2130555 0.05455614 0.1866084
#>             V27        V28       V29          V3       V30       V31        V32
#>  [1,] 0.7150026 0.80152091 0.8269567 0.050732555 0.8173406 0.8083878 0.69008344
#>  [2,] 0.5869147 0.78371674 0.7833787 0.027761955 0.5662552 0.2612908 0.31270858
#>  [3,] 0.6069358 0.74810304 0.7967870 0.098251439 0.7908240 0.8572303 0.85504449
#>  [4,] 0.7567000 0.80770000 0.8477000 0.144700000 0.9289000 0.9513000 0.79950000
#>  [5,] 0.3305041 0.25367790 0.2177294 0.074768588 0.2043479 0.1866031 0.23973381
#>  [6,] 0.9180622 0.94290990 0.9179078 0.010291826 0.9471229 0.9131438 0.80830821
#>  [7,] 0.9376704 0.87877169 0.7215040 0.040258236 0.5248007 0.3852944 0.35711305
#>  [8,] 0.9388871 0.95016827 0.8508501 0.033958661 0.7412532 0.6725136 0.52547174
#>  [9,] 0.8814159 0.85157711 0.5822575 0.039868550 0.3696731 0.3395164 0.45780591
#> [10,] 0.9417567 0.96311705 0.7937354 0.029546974 0.5671435 0.2967881 0.19148539
#> [11,] 0.1997057 0.09058063 0.2527969 0.012526881 0.3649171 0.3048758 0.45884875
#> [12,] 0.9108852 0.92132691 0.8660020 0.029570418 0.6833379 0.4777620 0.32802742
#> [13,] 0.7854291 0.67144534 0.6275674 0.077769368 0.4738213 0.2677764 0.15326507
#> [14,] 0.7601480 0.62638576 0.5161927 0.037414050 0.4663278 0.3556939 0.20794321
#> [15,] 0.7539000 0.60080000 0.5437000 0.062300000 0.5387000 0.5619000 0.51410000
#> [16,] 0.8150467 0.60516418 0.3174448 0.021213979 0.3622281 0.5486594 0.54049493
#> [17,] 0.2492982 0.37549466 0.5553235 0.028873490 0.7366557 0.7814180 0.74479959
#> [18,] 0.8274764 0.61060200 0.5168237 0.009469199 0.3615859 0.2682076 0.09500683
#> [19,] 0.7603207 0.68030328 0.7343234 0.007986611 0.8423751 0.6268247 0.42002587
#> [20,] 0.9026050 0.88081764 0.6976673 0.009416765 0.5136761 0.4330270 0.36221589
#> [21,] 0.7065646 0.81290808 0.8962590 0.022879089 0.8229928 0.7497492 0.70171112
#> [22,] 0.5380876 0.52433065 0.5118998 0.051396020 0.4627379 0.5263268 0.53428028
#> [23,] 0.8273907 0.76182269 0.6881711 0.037863216 0.6968140 0.5107467 0.20874690
#> [24,] 0.8324407 0.90094570 0.8995091 0.019766693 0.7863160 0.6542227 0.55802378
#> [25,] 0.4072356 0.66800614 0.8365428 0.010082087 0.7695349 0.5366574 0.19582961
#>              V33        V34        V35       V36        V37        V38
#>  [1,] 0.52027029 0.33431563 0.13469440 0.1693248 0.22660038 0.30100913
#>  [2,] 0.33093298 0.17913911 0.13279484 0.1150439 0.07046118 0.06254704
#>  [3,] 0.68311349 0.60865147 0.67021611 0.4235514 0.15303610 0.30385001
#>  [4,] 0.43620000 0.40480000 0.49520000 0.1712000 0.36520000 0.37630000
#>  [5,] 0.36039475 0.40394448 0.46066551 0.3055363 0.31175490 0.39924414
#>  [6,] 0.69379126 0.64481547 0.63423614 0.6845979 0.65748924 0.54636292
#>  [7,] 0.30241492 0.25411244 0.14734912 0.2144402 0.22452887 0.26154160
#>  [8,] 0.36965803 0.26661467 0.16495370 0.1063718 0.19208412 0.17622622
#>  [9,] 0.64676004 0.68790395 0.56573427 0.5172770 0.50299715 0.38080817
#> [10,] 0.25241040 0.47373322 0.71164832 0.8147022 0.78309425 0.67437233
#> [11,] 0.39541008 0.18937256 0.04926690 0.2640858 0.39646372 0.35265358
#> [12,] 0.19190293 0.06954762 0.06716363 0.1504066 0.20505004 0.26052699
#> [13,] 0.03342111 0.19889038 0.29242543 0.2383625 0.18976400 0.21975633
#> [14,] 0.16826191 0.24435061 0.36107159 0.4020804 0.32376386 0.24492905
#> [15,] 0.60840000 0.56210000 0.59560000 0.6078000 0.50250000 0.28290000
#> [16,] 0.37210234 0.40632196 0.55881928 0.5828977 0.53776885 0.49888012
#> [17,] 0.78983452 0.88723759 0.97676508 0.9256804 0.67886552 0.45615932
#> [18,] 0.12769677 0.15847900 0.20284735 0.1771879 0.04324289 0.09381188
#> [19,] 0.69051484 0.82635879 0.73554153 0.7294945 0.59562001 0.48588232
#> [20,] 0.26671931 0.14875220 0.18274490 0.2232904 0.24292499 0.26761509
#> [21,] 0.67782881 0.73902349 0.74377000 0.7030262 0.60551879 0.45537434
#> [22,] 0.53956895 0.52091201 0.54966660 0.5409298 0.37609093 0.28971518
#> [23,] 0.29163856 0.13482701 0.33808217 0.6310768 0.58772394 0.19400127
#> [24,] 0.54355925 0.56023536 0.51936787 0.4034963 0.19488729 0.13482997
#> [25,] 0.47781705 0.48726556 0.25301414 0.3642946 0.51348542 0.60380263
#>              V39         V4       V40       V41       V42        V43        V44
#>  [1,] 0.46443876 0.03510978 0.5309059 0.5222018 0.5017074 0.49887970 0.47917287
#>  [2,] 0.04682148 0.06319693 0.1022367 0.2495112 0.2901775 0.13093834 0.20716407
#>  [3,] 0.28277350 0.09431934 0.2111023 0.5981508 0.6582890 0.39979127 0.49690975
#>  [4,] 0.28410000 0.16440000 0.0427000 0.5331000 0.6952000 0.42880000 0.30630000
#>  [5,] 0.34809842 0.10638668 0.2726367 0.2738310 0.3224111 0.28003312 0.18580282
#>  [6,] 0.47294393 0.05160379 0.4878285 0.4815126 0.4410147 0.44371642 0.52038377
#>  [7,] 0.33772883 0.04045277 0.2816981 0.1416448 0.1999152 0.30453632 0.32903589
#>  [8,] 0.15226818 0.04846389 0.1796843 0.2311575 0.2175207 0.19462685 0.14732923
#>  [9,] 0.37047985 0.06189716 0.3702238 0.2571821 0.1586467 0.18806658 0.22611076
#> [10,] 0.56139674 0.03231897 0.3932340 0.2904251 0.2649409 0.33652606 0.37422181
#> [11,] 0.25954904 0.04064641 0.1730405 0.1490308 0.1097089 0.07665937 0.08259439
#> [12,] 0.25820495 0.02195049 0.2795185 0.2515159 0.2128934 0.23152331 0.19181860
#> [13,] 0.28502474 0.03691010 0.2593083 0.2843644 0.2363348 0.11502192 0.18183979
#> [14,] 0.34308114 0.09785244 0.3889344 0.4049129 0.4801894 0.35856293 0.21266888
#> [15,] 0.04770000 0.06000000 0.2811000 0.3422000 0.5147000 0.43720000 0.24700000
#> [16,] 0.57662105 0.01059867 0.6041840 0.4606121 0.2871344 0.22954980 0.16600586
#> [17,] 0.37508962 0.02216741 0.3400868 0.3520672 0.3628364 0.21071283 0.23431362
#> [18,] 0.17012678 0.01580026 0.2409937 0.3126312 0.2941709 0.19485104 0.08021855
#> [19,] 0.19940319 0.03374230 0.6882268 0.7915437 0.6617839 0.40728844 0.20348412
#> [20,] 0.21245235 0.01033622 0.1539845 0.1404726 0.1607028 0.15118297 0.13316777
#> [21,] 0.33462860 0.03001815 0.2877285 0.2305842 0.1719019 0.15641405 0.16640082
#> [22,] 0.21109499 0.05104696 0.3956299 0.3719449 0.4138068 0.38861682 0.24994978
#> [23,] 0.12134424 0.03341479 0.1905245 0.3834952 0.3917360 0.26669287 0.14107140
#> [24,] 0.17391780 0.02554591 0.1892133 0.1588473 0.1596782 0.22636020 0.17266670
#> [25,] 0.65906004 0.02015324 0.5044999 0.2302078 0.1858265 0.13588595 0.05565066
#>              V45         V46        V47        V48        V49         V5
#>  [1,] 0.61380854 0.473040079 0.33717236 0.25315170 0.14034531 0.07142095
#>  [2,] 0.21356549 0.179392898 0.10257509 0.05062558 0.04459108 0.10535466
#>  [3,] 0.61084032 0.561550477 0.35165224 0.16293299 0.06491078 0.05485935
#>  [4,] 0.58350000 0.569200000 0.26300000 0.11960000 0.09830000 0.16930000
#>  [5,] 0.13932088 0.065600204 0.07121488 0.05080995 0.02789491 0.10346327
#>  [6,] 0.49574298 0.333941598 0.23124071 0.19762672 0.10793310 0.08268653
#>  [7,] 0.30052555 0.203132434 0.13615763 0.09757396 0.05819148 0.06323039
#>  [8,] 0.13583301 0.122787833 0.12897796 0.11342871 0.06038551 0.05394466
#>  [9,] 0.15696119 0.120120958 0.08893230 0.08645270 0.03531643 0.05690990
#> [10,] 0.36998425 0.290753482 0.18012537 0.10295371 0.06467281 0.03808994
#> [11,] 0.05151863 0.038057184 0.05086826 0.04348271 0.01597033 0.04269690
#> [12,] 0.14450212 0.100793052 0.13221581 0.09088704 0.05940893 0.10824591
#> [13,] 0.16379675 0.139359722 0.10146707 0.08021886 0.05123697 0.09452696
#> [14,] 0.11074533 0.168824725 0.18995743 0.11136288 0.05578372 0.13889075
#> [15,] 0.17080000 0.134300000 0.08380000 0.07550000 0.03040000 0.13970000
#> [16,] 0.09207027 0.101140949 0.06768058 0.03700602 0.01530728 0.02290596
#> [17,] 0.17566479 0.144705519 0.14005649 0.10015396 0.04933262 0.04087355
#> [18,] 0.07639405 0.092393619 0.04562861 0.02213960 0.02473992 0.05234007
#> [19,] 0.26977185 0.401539342 0.27449214 0.08232174 0.04023406 0.07679968
#> [20,] 0.09297858 0.054701340 0.05637686 0.05260748 0.01539782 0.02988265
#> [21,] 0.15093341 0.118683565 0.07169612 0.05172498 0.02847819 0.04164535
#> [22,] 0.14482467 0.120470012 0.08237119 0.06945504 0.04185506 0.03892492
#> [23,] 0.16872535 0.179349427 0.12012411 0.11041448 0.05849071 0.01384766
#> [24,] 0.07768891 0.061827348 0.03994494 0.03586105 0.02223021 0.02712856
#> [25,] 0.01594658 0.003247672 0.07058648 0.07059958 0.01927087 0.04220137
#>               V50         V51         V52         V53         V54          V55
#>  [1,] 0.041493080 0.016101864 0.013140646 0.012760071 0.009094982  0.007046244
#>  [2,] 0.006300308 0.004818193 0.010783258 0.004683165 0.001984904 -0.002359370
#>  [3,] 0.032167803 0.045536057 0.042038333 0.021710875 0.026326113  0.025493328
#>  [4,] 0.037400000 0.029100000 0.015600000 0.019700000 0.013500000  0.012700000
#>  [5,] 0.018271196 0.019021941 0.019207939 0.021475491 0.015751380  0.014168817
#>  [6,] 0.028440400 0.016814025 0.013268934 0.011705058 0.007389612  0.004009036
#>  [7,] 0.017004422 0.020114513 0.011283677 0.008410232 0.008557028  0.007876394
#>  [8,] 0.014197472 0.015341458 0.013654662 0.007929447 0.007953566  0.005931353
#>  [9,] 0.021999717 0.008552408 0.012141359 0.005868555 0.018978755  0.011256375
#> [10,] 0.022900340 0.011211736 0.010677318 0.006510140 0.004123612  0.005400477
#> [11,] 0.012941893 0.013453128 0.009876496 0.001386980 0.005957298  0.003936090
#> [12,] 0.030057599 0.021451177 0.016599953 0.022212317 0.011100887  0.008362713
#> [13,] 0.019840611 0.025714892 0.019506934 0.018809381 0.026292951  0.021908657
#> [14,] 0.030424430 0.024189774 0.021306151 0.025604281 0.016885886  0.002756317
#> [15,] 0.007400000 0.006900000 0.002500000 0.010300000 0.007400000  0.012300000
#> [16,] 0.012293502 0.017371624 0.005553717 0.004322575 0.010198686  0.004824447
#> [17,] 0.013565470 0.016307228 0.005144613 0.007994489 0.006301593  0.010744109
#> [18,] 0.009300839 0.011751679 0.006667855 0.004834511 0.006496980  0.006853184
#> [19,] 0.056076242 0.041325205 0.033083774 0.015406536 0.011981490  0.011869545
#> [20,] 0.012725545 0.006559135 0.008992504 0.011706755 0.012819410  0.009052322
#> [21,] 0.016047394 0.009561274 0.009041362 0.006825463 0.006618671  0.007521612
#> [22,] 0.019739432 0.007248631 0.011734844 0.014088722 0.011488135  0.008832985
#> [23,] 0.023982342 0.011891929 0.012257440 0.010519734 0.010004826  0.009156116
#> [24,] 0.008436826 0.011738993 0.009781759 0.011563402 0.003603955  0.004366921
#> [25,] 0.010658718 0.005685011 0.019445149 0.026652120 0.016042099  0.020199196
#>               V56         V57         V58          V59         V6          V60
#>  [1,] 0.005266875 0.006823701 0.006375683 0.0063128368 0.12230627 0.0036877895
#>  [2,] 0.001859073 0.001883670 0.009901204 0.0034485603 0.14740970 0.0003557262
#>  [3,] 0.013136855 0.012644849 0.012112197 0.0241851813 0.12780534 0.0133734774
#>  [4,] 0.013800000 0.013300000 0.013100000 0.0154000000 0.08440000 0.0218000000
#>  [5,] 0.017186480 0.009170530 0.011366715 0.0197694581 0.09708187 0.0138859707
#>  [6,] 0.005866389 0.006263673 0.004213601 0.0009321086 0.11814986 0.0023536250
#>  [7,] 0.006037633 0.005511795 0.005490348 0.0052590633 0.10156731 0.0049976433
#>  [8,] 0.006837269 0.005591684 0.006445516 0.0049138963 0.09721663 0.0054051328
#>  [9,] 0.012763174 0.008172522 0.015665440 0.0162991518 0.07567363 0.0086314455
#> [10,] 0.005202382 0.004702063 0.003433871 0.0047018723 0.05532385 0.0043031744
#> [11,] 0.004216110 0.007368743 0.005315927 0.0018157219 0.06607462 0.0022575636
#> [12,] 0.008757106 0.012538777 0.009993232 0.0105783121 0.14580802 0.0080644472
#> [13,] 0.010523428 0.012422488 0.015985131 0.0064781655 0.18939446 0.0062676306
#> [14,] 0.008371153 0.012658055 0.009606003 0.0111691927 0.06559528 0.0084921772
#> [15,] 0.006900000 0.007600000 0.007300000 0.0030000000 0.18830000 0.0138000000
#> [16,] 0.008264312 0.005442781 0.001286043 0.0045245877 0.09014319 0.0042562680
#> [17,] 0.007734615 0.007708889 0.011874536 0.0083322817 0.07922673 0.0024524601
#> [18,] 0.003979999 0.001617346 0.001181342 0.0032079323 0.05272188 0.0021743129
#> [19,] 0.002578207 0.006871855 0.001955964 0.0055132332 0.11601104 0.0066860005
#> [20,] 0.007277167 0.004557985 0.006351509 0.0041655843 0.05277130 0.0040117067
#> [21,] 0.005060582 0.008167851 0.005278725 0.0066490273 0.09631450 0.0057760087
#> [22,] 0.012148443 0.007601137 0.004237161 0.0060205784 0.07831119 0.0114905685
#> [23,] 0.007904850 0.009340371 0.004459234 0.0070432459 0.05573710 0.0028221920
#> [24,] 0.003877746 0.009740097 0.005295405 0.0047688393 0.05262327 0.0060255371
#> [25,] 0.005132228 0.002479314 0.004109272 0.0091059363 0.05553858 0.0090836220
#>               V7         V8         V9
#>  [1,] 0.11768729 0.13800205 0.26150475
#>  [2,] 0.15592579 0.20563569 0.25761010
#>  [3,] 0.07712715 0.09849585 0.21432234
#>  [4,] 0.07150000 0.09470000 0.15830000
#>  [5,] 0.11105786 0.07277884 0.06785239
#>  [6,] 0.15103670 0.14380474 0.30796678
#>  [7,] 0.12269973 0.15819163 0.24985721
#>  [8,] 0.11044398 0.11739626 0.17018000
#>  [9,] 0.20691954 0.34899973 0.53087707
#> [10,] 0.05487912 0.04572875 0.08116094
#> [11,] 0.12510935 0.13261112 0.19885213
#> [12,] 0.13163635 0.17267098 0.20301763
#> [13,] 0.20484491 0.25881185 0.35951016
#> [14,] 0.11393998 0.11188202 0.03326512
#> [15,] 0.14220000 0.14470000 0.04870000
#> [16,] 0.08340686 0.06447627 0.07792866
#> [17,] 0.08792485 0.03017058 0.02711649
#> [18,] 0.05620399 0.04039794 0.09916868
#> [19,] 0.16913620 0.12733830 0.05142305
#> [20,] 0.07493851 0.06365775 0.07879525
#> [21,] 0.10747612 0.09194763 0.07954050
#> [22,] 0.11501512 0.14974288 0.08710620
#> [23,] 0.11603346 0.15592056 0.14637066
#> [24,] 0.02758699 0.04970280 0.07387361
#> [25,] 0.02477947 0.05670534 0.13997939
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406