Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.015218055 0.31363252 0.40601163 0.50101967 0.49670593 0.41930562
#>  [2,] 0.047221021 0.36750717 0.46312406 0.54032437 0.60405454 0.70048602
#>  [3,] 0.023881094 0.20699729 0.26311415 0.28846372 0.28290929 0.35268369
#>  [4,] 0.022888687 0.21898187 0.27648685 0.31810953 0.24280243 0.18295295
#>  [5,] 0.027245838 0.23156236 0.25566170 0.29971986 0.28556419 0.24030934
#>  [6,] 0.082988319 0.57133055 0.45214268 0.19580100 0.21327769 0.32664644
#>  [7,] 0.024631243 0.11801595 0.20768863 0.25168940 0.23038045 0.14298810
#>  [8,] 0.094646763 0.25231749 0.39353747 0.46484880 0.61349014 0.70698414
#>  [9,] 0.024743041 0.15386310 0.30170420 0.46540967 0.55813179 0.55340141
#> [10,] 0.027391467 0.22551707 0.28942257 0.30362279 0.26375333 0.18888244
#> [11,] 0.048370125 0.01757083 0.19428713 0.21008301 0.16039659 0.31108164
#> [12,] 0.036124951 0.22981816 0.28226968 0.27986792 0.30873544 0.34772554
#> [13,] 0.041246843 0.24735632 0.26830806 0.28097848 0.33666485 0.31518152
#> [14,] 0.009487929 0.08011116 0.09153893 0.10461894 0.15841286 0.18630153
#> [15,] 0.012448319 0.02537433 0.17053633 0.37614649 0.28812386 0.24703431
#> [16,] 0.018029238 0.12909161 0.14025196 0.13798068 0.14601138 0.15160235
#> [17,] 0.022760242 0.08193542 0.08224788 0.14101545 0.18745538 0.14550201
#> [18,] 0.011633599 0.12281646 0.11221786 0.03169730 0.06691060 0.05635752
#> [19,] 0.031700000 0.35130000 0.17860000 0.06580000 0.05130000 0.37520000
#> [20,] 0.063864060 0.28101715 0.26839940 0.10023895 0.13315993 0.46692974
#> [21,] 0.022114635 0.17279334 0.16058705 0.18979740 0.30723290 0.29262214
#> [22,] 0.029133613 0.10468811 0.24439780 0.37092741 0.30518841 0.27159487
#> [23,] 0.018796734 0.11611233 0.13577125 0.09274322 0.09480307 0.10222638
#> [24,] 0.025177870 0.16990614 0.20915759 0.10165645 0.18494475 0.37917742
#> [25,] 0.020939253 0.09920141 0.09614576 0.10151993 0.19688662 0.34859196
#>              V15        V16       V17       V18       V19         V2       V20
#>  [1,] 0.34483423 0.33393926 0.3141451 0.2997444 0.4258150 0.03318225 0.6935917
#>  [2,] 0.75765276 0.81490992 0.9580883 0.9172454 0.9001109 0.04580624 0.8278975
#>  [3,] 0.40655097 0.48912116 0.5495491 0.6247775 0.6790882 0.02289383 0.7381773
#>  [4,] 0.13699414 0.17521486 0.1818111 0.2827873 0.4817067 0.02974860 0.5993877
#>  [5,] 0.15904086 0.16806409 0.2254204 0.3282424 0.5087830 0.03258172 0.6556167
#>  [6,] 0.51035525 0.52613633 0.1887500 0.1532988 0.3618784 0.09290231 0.8847249
#>  [7,] 0.09351704 0.17276380 0.2571880 0.3432078 0.4365670 0.04327622 0.5572798
#>  [8,] 0.73390075 0.79050580 0.8394011 0.8616022 0.9356809 0.09108889 0.8762070
#>  [9,] 0.63263230 0.74335306 0.9007041 0.9821300 0.9495621 0.04453692 0.8761437
#> [10,] 0.11000797 0.11463573 0.1877597 0.2404058 0.2588041 0.03879550 0.2179062
#> [11,] 0.38244625 0.46100556 0.4137311 0.3497159 0.4010245 0.08856172 0.2607942
#> [12,] 0.46944782 0.50777845 0.6103833 0.6676083 0.7679945 0.05312441 0.8346866
#> [13,] 0.26976754 0.29597257 0.3309600 0.2531444 0.3712626 0.05691895 0.4819555
#> [14,] 0.16315038 0.16701926 0.1723350 0.1976269 0.2202128 0.01173435 0.2254483
#> [15,] 0.07705627 0.06625278 0.1665619 0.4202389 0.3076218 0.05151523 0.6222169
#> [16,] 0.19556135 0.24824248 0.2815601 0.2879873 0.2217836 0.02079817 0.2107595
#> [17,] 0.19913360 0.20987161 0.2232818 0.2456214 0.2633618 0.04049152 0.2501507
#> [18,] 0.09391141 0.05766441 0.1016791 0.2201023 0.3539682 0.01329338 0.3788461
#> [19,] 0.54190000 0.54400000 0.5150000 0.4262000 0.2024000 0.09560000 0.4233000
#> [20,] 0.67350361 0.68712554 0.7269739 0.5894727 0.3988870 0.04093855 0.2782169
#> [21,] 0.27974337 0.37062183 0.3463005 0.3812310 0.6224122 0.03503284 0.7661215
#> [22,] 0.30181270 0.45142681 0.5992783 0.5388868 0.4897780 0.05417764 0.5875494
#> [23,] 0.09460537 0.16911900 0.2607448 0.3051749 0.3204188 0.01146199 0.4138984
#> [24,] 0.48648814 0.68088591 0.8680444 0.9178161 0.7744737 0.02560916 0.5019255
#> [25,] 0.50176826 0.65674045 0.7603275 0.7855188 0.8087706 0.02029697 0.8287394
#>             V21       V22       V23       V24       V25       V26        V27
#>  [1,] 0.7799735 0.7893476 0.6875372 0.7155397 0.7798253 0.8929305 0.99748241
#>  [2,] 0.7933421 0.6651464 0.3812304 0.1546263 0.2127430 0.2455455 0.32967719
#>  [3,] 0.8474795 0.9397888 0.9225237 0.8555446 0.7882025 0.7825203 0.79652258
#>  [4,] 0.7487641 0.8035582 0.7730807 0.8696728 0.8953215 0.9618065 1.00005633
#>  [5,] 0.8011794 0.8346928 0.8346869 0.9422296 0.9710842 0.9896109 0.99705273
#>  [6,] 0.9128146 0.3862298 0.1489060 0.1515081 0.2038227 0.3733071 0.27096289
#>  [7,] 0.6276369 0.7070595 0.7379253 0.6807208 0.7240615 0.8371555 0.95463264
#>  [8,] 0.8728741 0.8574228 0.8786844 0.7834446 0.5564277 0.5433129 0.48350591
#>  [9,] 0.7612836 0.6641001 0.5931008 0.5402772 0.3421341 0.1065528 0.07352811
#> [10,] 0.3177784 0.4672359 0.6142607 0.7255558 0.7792469 0.8242070 0.88670275
#> [11,] 0.2009997 0.1244561 0.3074072 0.5082051 0.4051524 0.3815760 0.45504812
#> [12,] 0.9265345 0.9847675 0.9113672 0.7993487 0.6559129 0.5667177 0.47480605
#> [13,] 0.5349723 0.5431738 0.5517084 0.5553752 0.5781972 0.6333260 0.68184466
#> [14,] 0.2634502 0.3707661 0.4628191 0.5922722 0.6425477 0.6895452 0.71549718
#> [15,] 0.7440357 0.7557146 0.7756781 0.8113534 1.0178883 0.9579999 0.75974330
#> [16,] 0.3315982 0.3584425 0.4408547 0.6377109 0.8561746 0.8874793 0.75865305
#> [17,] 0.2927724 0.2716805 0.2471687 0.2612833 0.2723043 0.3510410 0.36457428
#> [18,] 0.4579820 0.7886953 0.8880439 0.8526610 0.7302958 0.7399815 0.62708926
#> [19,] 0.7723000 0.9735000 0.9390000 0.5559000 0.5268000 0.6826000 0.57130000
#> [20,] 0.5016597 0.7495781 0.8566178 0.9540437 0.9017000 0.6498563 0.64993701
#> [21,] 0.6570963 0.4886909 0.5163544 0.5621398 0.6051662 0.7050678 0.76328065
#> [22,] 0.3947253 0.2460381 0.3804892 0.8097952 0.8565086 0.5319491 0.41257615
#> [23,] 0.5458262 0.6680367 0.7676136 0.7823357 0.7844152 0.8714925 0.89552952
#> [24,] 0.5398019 0.6987048 0.7859783 0.8244222 0.7230608 0.6425015 0.56291047
#> [25,] 0.9064942 0.9113461 0.8923921 0.8920640 0.8774729 0.9141912 0.85951583
#>             V28       V29         V3       V30        V31       V32       V33
#>  [1,] 0.9352680 0.7874001 0.05521609 0.5927023 0.53980180 0.6156711 0.4992286
#>  [2,] 0.3023058 0.1132724 0.05291526 0.1234705 0.06941621 0.1264037 0.2489155
#>  [3,] 0.7411663 0.6209956 0.03258812 0.4976336 0.36082612 0.3069792 0.2284510
#>  [4,] 0.9555736 0.8362384 0.04269576 0.6352774 0.46075450 0.3321392 0.2857017
#>  [5,] 0.9437030 0.8276651 0.04351451 0.6134970 0.41827836 0.3338547 0.2410263
#>  [6,] 0.5592361 0.5372010 0.06126475 0.7231075 0.36291721 0.3985327 0.6305268
#>  [7,] 0.9453064 0.7547013 0.04843683 0.5851899 0.30881899 0.2330733 0.2688579
#>  [8,] 0.5294923 0.4472217 0.10024542 0.2992273 0.23114934 0.1777350 0.1996948
#>  [9,] 0.1068771 0.1242621 0.04642188 0.1705959 0.31640967 0.4308873 0.3830453
#> [10,] 0.9620919 0.8872117 0.03285539 0.6188243 0.58933371 0.5149819 0.4103490
#> [11,] 0.4797812 0.4670897 0.12186741 0.4356106 0.40782473 0.3926479 0.4681622
#> [12,] 0.2817832 0.2658365 0.07731891 0.3194506 0.22741672 0.1855788 0.3376850
#> [13,] 0.7806001 0.7976066 0.06901857 0.8010926 0.80899830 0.7559375 0.6058840
#> [14,] 0.8371337 0.9546126 0.01907292 0.9193336 0.78760887 0.7058700 0.7142428
#> [15,] 0.6204075 0.5944940 0.06902064 0.5563669 0.55493336 0.5166713 0.6098096
#> [16,] 0.6887037 0.5812947 0.01756574 0.3916502 0.28451803 0.1922168 0.1780840
#> [17,] 0.5211351 0.6611573 0.02577915 0.7312817 0.70798891 0.6430264 0.6955603
#> [18,] 0.7015920 0.8264614 0.03446158 0.8048429 0.65444702 0.6842341 0.8414972
#> [19,] 0.5429000 0.2177000 0.13210000 0.2149000 0.58110000 0.6323000 0.2965000
#> [20,] 0.6538694 0.4986959 0.05344533 0.5670986 0.47150777 0.2171921 0.2430689
#> [21,] 0.6313619 0.4224883 0.04446336 0.3049996 0.30980037 0.3862562 0.3573684
#> [22,] 0.6988155 0.8201387 0.05811730 0.5386716 0.58723812 0.7005424 0.6231262
#> [23,] 0.8744388 0.8736465 0.01455834 0.7801326 0.69073571 0.6270167 0.5881930
#> [24,] 0.4286916 0.3765450 0.06170255 0.4387138 0.24781965 0.1731329 0.1706237
#> [25,] 0.7159896 0.5418732 0.02493053 0.4817760 0.39432044 0.3091915 0.2619726
#>             V34        V35        V36        V37       V38        V39
#>  [1,] 0.2871377 0.04190187 0.31608892 0.33930881 0.3939331 0.54961483
#>  [2,] 0.1517924 0.26684096 0.00220438 0.15111320 0.2297458 0.17712157
#>  [3,] 0.1974988 0.19468735 0.18152766 0.21647663 0.2646417 0.29491451
#>  [4,] 0.1883061 0.10979924 0.01464758 0.06043429 0.1165986 0.09950644
#>  [5,] 0.2287213 0.19724401 0.09848580 0.12446604 0.2036039 0.15755849
#>  [6,] 0.7822967 0.87218618 0.85648986 0.83991823 0.8937273 0.72319857
#>  [7,] 0.4990729 0.72245978 0.81794688 0.74368060 0.6286804 0.50444796
#>  [8,] 0.2138019 0.23793058 0.15377513 0.18407656 0.2845573 0.36545061
#>  [9,] 0.4927583 0.48758157 0.33708941 0.21907539 0.3748386 0.35799159
#> [10,] 0.2374597 0.13236574 0.07681148 0.16548908 0.1093935 0.15531268
#> [11,] 0.4255293 0.28023852 0.22439527 0.11810488 0.2393360 0.33119111
#> [12,] 0.4135193 0.38005431 0.38159913 0.42508650 0.3442616 0.31760890
#> [13,] 0.4973234 0.41094530 0.29085981 0.25607915 0.3476879 0.40379197
#> [14,] 0.6057015 0.52446088 0.45959421 0.37460291 0.3028894 0.27028839
#> [15,] 0.5517742 0.59817811 0.63615155 0.53181334 0.2869881 0.02322426
#> [16,] 0.1938430 0.23726723 0.21004015 0.13024945 0.1394485 0.16215961
#> [17,] 0.7695463 0.83474002 0.84384848 0.74072999 0.6694238 0.59470630
#> [18,] 0.9162622 0.85335386 0.64647889 0.38145034 0.2234952 0.22694460
#> [19,] 0.1873000 0.29690000 0.51630000 0.61530000 0.4283000 0.54790000
#> [20,] 0.1623983 0.49799157 0.72188414 0.74377290 0.3839824 0.14122704
#> [21,] 0.3307841 0.37303043 0.41274825 0.43611443 0.3915290 0.46428441
#> [22,] 0.8682256 0.82358546 0.77904704 0.80993390 0.5728908 0.42720829
#> [23,] 0.6127693 0.63148360 0.59933414 0.51811234 0.4158475 0.29840727
#> [24,] 0.1161912 0.37040433 0.51245112 0.40409439 0.2193438 0.12530094
#> [25,] 0.2711324 0.20393433 0.12292322 0.16558248 0.1381549 0.07803743
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.05108713 0.43894506 0.23622533 0.42108979 0.45413080 0.48913168
#>  [2,] 0.13253250 0.05767885 0.05735153 0.24920271 0.22710110 0.02718147
#>  [3,] 0.06623462 0.30232719 0.27750574 0.24464102 0.24563089 0.19987331
#>  [4,] 0.04410186 0.10411581 0.03535967 0.11754534 0.20479529 0.21128474
#>  [5,] 0.02495413 0.15345374 0.04933064 0.10273556 0.20992307 0.23265426
#>  [6,] 0.10620376 0.49754544 0.34220957 0.31117936 0.25933432 0.28985098
#>  [7,] 0.06031343 0.36830536 0.35584093 0.35911267 0.39011776 0.34064333
#>  [8,] 0.12506810 0.29251844 0.15112399 0.18838833 0.31798562 0.18821637
#>  [9,] 0.06095261 0.15950284 0.07829914 0.12036359 0.13968182 0.04463494
#> [10,] 0.06209968 0.20366538 0.26605526 0.30116699 0.21437451 0.10780445
#> [11,] 0.17300273 0.34954845 0.37464927 0.41070677 0.32508663 0.28239165
#> [12,] 0.08584327 0.27756876 0.23623105 0.31937026 0.26568015 0.15318185
#> [13,] 0.06727808 0.36246381 0.54870112 0.58987082 0.46624214 0.44036850
#> [14,] 0.02809746 0.22618054 0.15882611 0.11954134 0.16866964 0.17800978
#> [15,] 0.06259599 0.29107823 0.36658835 0.55138597 0.45578259 0.24958269
#> [16,] 0.02494261 0.19977643 0.24960857 0.22028789 0.16420204 0.11079557
#> [17,] 0.01101781 0.56424319 0.51959127 0.47303532 0.34129022 0.23423116
#> [18,] 0.02901369 0.33945183 0.32486357 0.25351928 0.30902140 0.24017443
#> [19,] 0.14080000 0.61330000 0.50170000 0.23770000 0.19570000 0.17490000
#> [20,] 0.03518821 0.36894881 0.45337622 0.41285270 0.28376847 0.20314215
#> [21,] 0.03121664 0.44814522 0.28301761 0.22758528 0.18363905 0.16298305
#> [22,] 0.07524415 0.38621112 0.32070204 0.29540492 0.10340239 0.17999118
#> [23,] 0.02448362 0.24995330 0.21430201 0.14026687 0.14310301 0.12919609
#> [24,] 0.06485239 0.30156445 0.38427613 0.28673260 0.16897812 0.17245065
#> [25,] 0.03919556 0.10677821 0.09910201 0.06414574 0.06573525 0.08996981
#>              V45        V46        V47        V48        V49         V5
#>  [1,] 0.51038026 0.23275297 0.14555498 0.16444900 0.08056680 0.07853209
#>  [2,] 0.17245102 0.12021780 0.04602666 0.06408525 0.04454030 0.19952566
#>  [3,] 0.15949011 0.15638351 0.13892572 0.08409492 0.05722879 0.09992205
#>  [4,] 0.19493255 0.20200098 0.17069625 0.10825631 0.04920128 0.07374183
#>  [5,] 0.18218478 0.19665666 0.15739099 0.08565064 0.05367388 0.06520831
#>  [6,] 0.42712269 0.31485961 0.13444897 0.12328889 0.06445791 0.08339517
#>  [7,] 0.34677237 0.25955737 0.13001139 0.06113500 0.05438457 0.04542394
#>  [8,] 0.07223083 0.07716682 0.05519950 0.05335548 0.05237707 0.15265990
#>  [9,] 0.05895661 0.04651114 0.04282274 0.04071277 0.03035874 0.08040011
#> [10,] 0.13456183 0.18349895 0.21594909 0.17165825 0.08020503 0.08879507
#> [11,] 0.33587203 0.31222158 0.19723699 0.16376106 0.08656576 0.13562806
#> [12,] 0.11068559 0.06417840 0.06644400 0.05763369 0.03503145 0.12232969
#> [13,] 0.55907546 0.49470845 0.32749324 0.21041965 0.11561269 0.08478937
#> [14,] 0.14490623 0.08506111 0.04597944 0.04357322 0.02114039 0.02585293
#> [15,] 0.16720746 0.13115007 0.07069967 0.06717043 0.02480783 0.13371032
#> [16,] 0.10309964 0.09929201 0.04770321 0.03997233 0.02801217 0.04688025
#> [17,] 0.21421648 0.19968476 0.16490726 0.10968151 0.06172860 0.02981037
#> [18,] 0.14741077 0.08862058 0.05496139 0.03951703 0.01784987 0.05532741
#> [19,] 0.13040000 0.05970000 0.11240000 0.10470000 0.05070000 0.16740000
#> [20,] 0.24002641 0.25003976 0.21970090 0.18758612 0.12506452 0.03343219
#> [21,] 0.14555190 0.13757593 0.08410054 0.05492226 0.04257251 0.05280056
#> [22,] 0.15522706 0.16549319 0.12277808 0.07793615 0.04084695 0.12430162
#> [23,] 0.09004915 0.07227804 0.04476439 0.03653895 0.01728775 0.05065468
#> [24,] 0.16600561 0.06735371 0.10640793 0.11486773 0.06524303 0.11351425
#> [25,] 0.08838032 0.05207505 0.07063360 0.04948787 0.02938875 0.06190982
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.014313463 0.026357795 0.019080826 0.007438520 0.012346534 0.013165860
#>  [2,] 0.044442799 0.031912375 0.030441496 0.028412618 0.015565106 0.021544701
#>  [3,] 0.025692066 0.016526435 0.015318981 0.013071959 0.014668956 0.006502539
#>  [4,] 0.018911730 0.017712877 0.010046211 0.007123893 0.008712604 0.003382180
#>  [5,] 0.017445087 0.018210078 0.012238650 0.003213235 0.003602106 0.005621204
#>  [6,] 0.027469650 0.023546281 0.023993611 0.017597331 0.024923868 0.020011977
#>  [7,] 0.023196915 0.008615365 0.012904058 0.008907751 0.003623106 0.007965369
#>  [8,] 0.020646856 0.021447863 0.014219167 0.012557700 0.026234843 0.016021596
#>  [9,] 0.011479039 0.015267899 0.012326331 0.008782937 0.020305417 0.018995949
#> [10,] 0.016136271 0.024271659 0.015798002 0.008363597 0.009229115 0.004879067
#> [11,] 0.023289940 0.021757964 0.034174663 0.024195786 0.019144225 0.028609842
#> [12,] 0.010012446 0.015096889 0.013540528 0.007752475 0.019359834 0.008505061
#> [13,] 0.042476278 0.034100844 0.025719208 0.016295436 0.015440962 0.014458830
#> [14,] 0.011137996 0.009424675 0.007767572 0.009951993 0.007220361 0.005991338
#> [15,] 0.005541059 0.004462110 0.002525371 0.010763211 0.007901225 0.012902982
#> [16,] 0.012051612 0.009234737 0.009878900 0.007063164 0.009461381 0.008531063
#> [17,] 0.026966038 0.019580812 0.010293094 0.013245999 0.009381964 0.013678466
#> [18,] 0.015046527 0.012316308 0.013545339 0.016847382 0.012980287 0.005641930
#> [19,] 0.015900000 0.019500000 0.020100000 0.024800000 0.013100000 0.007000000
#> [20,] 0.032701503 0.015324521 0.014312771 0.010314472 0.008688653 0.004988687
#> [21,] 0.020980157 0.011809392 0.009406056 0.006765339 0.017134326 0.016170205
#> [22,] 0.010841647 0.018332372 0.007769807 0.014404077 0.006535062 0.017012343
#> [23,] 0.010988814 0.003903344 0.005440911 0.004737671 0.005437243 0.006608993
#> [24,] 0.006555471 0.022446829 0.010360563 0.009119853 0.008481934 0.006163354
#> [25,] 0.017143424 0.011650171 0.008478370 0.006914285 0.008442752 0.005658588
#>               V56         V57         V58         V59         V6         V60
#>  [1,] 0.002684568 0.007960525 0.002848045 0.004933794 0.09597145 0.002475118
#>  [2,] 0.009433384 0.005815229 0.021892878 0.017382308 0.21152154 0.014385898
#>  [3,] 0.010228093 0.007955600 0.009789896 0.008297930 0.11007301 0.006763005
#>  [4,] 0.002364526 0.003166926 0.002007112 0.003327507 0.08622217 0.005336272
#>  [5,] 0.002590278 0.002962625 0.003899288 0.003079244 0.04721966 0.003543689
#>  [6,] 0.016932257 0.012441077 0.014961353 0.014084514 0.11353396 0.013271767
#>  [7,] 0.006092192 0.003882613 0.004152042 0.005727258 0.06865794 0.004941508
#>  [8,] 0.010844704 0.007086073 0.012098060 0.015455351 0.14789907 0.007414806
#>  [9,] 0.008364861 0.003560304 0.010020101 0.008418229 0.04898901 0.004010634
#> [10,] 0.007026270 0.006890664 0.008993484 0.007861478 0.14785526 0.004662353
#> [11,] 0.007471194 0.024131556 0.023138946 0.020481852 0.08763419 0.012488349
#> [12,] 0.008614794 0.012184772 0.017577381 0.008179573 0.13413761 0.004824278
#> [13,] 0.009982785 0.010619802 0.011833352 0.012415604 0.10797283 0.008119616
#> [14,] 0.006378324 0.005496549 0.003007489 0.007612238 0.04572349 0.004484915
#> [15,] 0.008855393 0.008340702 0.006286965 0.003721638 0.18295199 0.015477163
#> [16,] 0.004810390 0.004237935 0.003691080 0.004257512 0.05920943 0.003485049
#> [17,] 0.008525960 0.004299455 0.010190584 0.008320089 0.07465705 0.004663746
#> [18,] 0.010540156 0.017308696 0.007191288 0.007223138 0.05557168 0.010693350
#> [19,] 0.013800000 0.009200000 0.014300000 0.003600000 0.17100000 0.010300000
#> [20,] 0.004754291 0.020947726 0.008895195 0.005567221 0.05887952 0.004601684
#> [21,] 0.008188342 0.011664135 0.012220056 0.008982616 0.06029359 0.005848188
#> [22,] 0.007450044 0.012347867 0.007309689 0.016332239 0.19252675 0.008099258
#> [23,] 0.006314902 0.006253285 0.005253990 0.003121009 0.08554556 0.003439556
#> [24,] 0.007038678 0.008719853 0.013534614 0.018440978 0.11104830 0.015294663
#> [25,] 0.009285203 0.007169136 0.003942979 0.004829267 0.08583602 0.004121666
#>               V7         V8         V9
#>  [1,] 0.10376444 0.18063456 0.26360991
#>  [2,] 0.18980108 0.10669193 0.18960451
#>  [3,] 0.12801990 0.13096291 0.17578074
#>  [4,] 0.09254976 0.11417787 0.16992852
#>  [5,] 0.09740954 0.11332157 0.17649551
#>  [6,] 0.22314844 0.42093632 0.60118702
#>  [7,] 0.05828795 0.08653325 0.10725859
#>  [8,] 0.19042208 0.23984301 0.21493685
#>  [9,] 0.10109732 0.12440749 0.07535838
#> [10,] 0.13249971 0.12613566 0.15190780
#> [11,] 0.09850078 0.17193176 0.09005132
#> [12,] 0.10759692 0.12181461 0.17778216
#> [13,] 0.09803539 0.12273246 0.22779130
#> [14,] 0.05309891 0.04804113 0.06833647
#> [15,] 0.13839997 0.14959440 0.01943457
#> [16,] 0.07457063 0.07421983 0.10367572
#> [17,] 0.09857449 0.09361963 0.02609380
#> [18,] 0.08956349 0.07180221 0.03446492
#> [19,] 0.07310000 0.14010000 0.20830000
#> [20,] 0.17447309 0.27295175 0.27136626
#> [21,] 0.09165553 0.06511559 0.11302976
#> [22,] 0.18978776 0.16678872 0.13989096
#> [23,] 0.08511890 0.09359302 0.11324747
#> [24,] 0.17969848 0.16319062 0.13404558
#> [25,] 0.11572416 0.09134756 0.11312485
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551