Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13        V14
#>  [1,] 0.02790028 0.24357582 0.30236272 0.31219593 0.23058378 0.12277396
#>  [2,] 0.02522891 0.24676521 0.28645205 0.31454830 0.30823703 0.32101276
#>  [3,] 0.04830904 0.07128911 0.16964612 0.28748704 0.41019369 0.51779629
#>  [4,] 0.01980745 0.07901707 0.12598812 0.09923656 0.09789198 0.21177802
#>  [5,] 0.02580000 0.29540000 0.25060000 0.26010000 0.22490000 0.21150000
#>  [6,] 0.01143428 0.31730842 0.32442090 0.34314478 0.24859460 0.17880866
#>  [7,] 0.01083122 0.24769695 0.27579717 0.35593661 0.32130997 0.24938330
#>  [8,] 0.04377563 0.40559976 0.52544960 0.53174336 0.62941678 0.60795496
#>  [9,] 0.02256355 0.29904415 0.22761201 0.30610843 0.28684301 0.20871256
#> [10,] 0.00470000 0.34690000 0.32650000 0.32630000 0.23010000 0.12530000
#> [11,] 0.03022047 0.37926519 0.46738765 0.51682715 0.51777562 0.39964227
#> [12,] 0.05683489 0.24616607 0.34983536 0.52051683 0.60649870 0.76129617
#> [13,] 0.05391361 0.18789793 0.22046590 0.22885717 0.31835816 0.34685050
#> [14,] 0.02469243 0.26180293 0.25373729 0.10581639 0.20917040 0.49098548
#> [15,] 0.01010862 0.10630522 0.18726427 0.14754758 0.16439043 0.25418314
#> [16,] 0.05726027 0.31336109 0.25541027 0.09931725 0.15812491 0.46622442
#> [17,] 0.02433763 0.10710308 0.14173595 0.13895319 0.21793976 0.35273210
#> [18,] 0.01767035 0.12547292 0.13394041 0.08912682 0.10546281 0.08996356
#> [19,] 0.04160947 0.07441141 0.17621639 0.36532044 0.27975746 0.17921179
#> [20,] 0.02304200 0.62027701 0.63725807 0.72449793 0.56648577 0.54212020
#> [21,] 0.01663471 0.10727672 0.14152597 0.14757305 0.18557360 0.22912156
#> [22,] 0.02614324 0.00389470 0.02190896 0.19064309 0.12814558 0.08881285
#> [23,] 0.02430715 0.04094306 0.12088990 0.32254007 0.33916434 0.31234067
#> [24,] 0.01676147 0.09132190 0.09650234 0.10380825 0.15014157 0.18559079
#> [25,] 0.02541642 0.12644804 0.14586673 0.15216476 0.13924031 0.14861799
#>              V15        V16       V17       V18       V19          V2
#>  [1,] 0.14218566 0.18120675 0.2300616 0.2388669 0.2902164 0.045796772
#>  [2,] 0.37120910 0.45603618 0.5466408 0.6614168 0.7227651 0.025327220
#>  [3,] 0.59473250 0.59252111 0.6922304 0.8101121 0.9405067 0.063688582
#>  [4,] 0.28447845 0.46731111 0.5095450 0.6239927 0.6968832 0.006855418
#>  [5,] 0.12700000 0.11930000 0.1794000 0.2185000 0.1646000 0.043300000
#>  [6,] 0.11559787 0.23630731 0.3027189 0.3613641 0.2595016 0.037829019
#>  [7,] 0.17301558 0.20474360 0.2302878 0.3639635 0.6214861 0.013311864
#>  [8,] 0.67020793 0.71727092 0.8706975 0.9462825 0.8797890 0.048759748
#>  [9,] 0.09069939 0.14633810 0.1513126 0.2073935 0.1871431 0.020606577
#> [10,] 0.21020000 0.24010000 0.1928000 0.1673000 0.1228000 0.005900000
#> [11,] 0.28921899 0.30578205 0.3280382 0.3211217 0.4715214 0.039038274
#> [12,] 0.87310562 0.92434578 0.9811498 0.9306612 0.9400516 0.060176589
#> [13,] 0.31221289 0.37708635 0.4130439 0.3077395 0.4063737 0.078568083
#> [14,] 0.66147632 0.74210442 0.8056708 0.6755000 0.4696101 0.061850354
#> [15,] 0.31190047 0.27608685 0.1930727 0.1712967 0.2047077 0.019883402
#> [16,] 0.64247932 0.71358295 0.7388152 0.6297038 0.4226410 0.055580606
#> [17,] 0.58126683 0.80943890 0.8753869 0.8444978 0.8742671 0.021207736
#> [18,] 0.08408094 0.09364999 0.1604859 0.2478512 0.3229986 0.013090886
#> [19,] 0.17337508 0.29893432 0.5007235 0.6458057 0.5935343 0.038012132
#> [20,] 0.64725763 0.69425011 0.6954596 0.7809256 0.9105863 0.056322897
#> [21,] 0.27110712 0.29382381 0.2477889 0.2326676 0.2488904 0.021080907
#> [22,] 0.09018950 0.10922240 0.1406422 0.1315450 0.1379333 0.036232690
#> [23,] 0.05969336 0.09090885 0.2163579 0.4089654 0.3882198 0.026006220
#> [24,] 0.22070099 0.24627281 0.2762283 0.3677221 0.4349449 0.026242360
#> [25,] 0.21037396 0.47981376 0.6838320 0.5657438 0.2933837 0.015292746
#>              V20        V21        V22        V23       V24       V25       V26
#>  [1,] 0.29307243 0.42405386 0.54797832 0.63150878 0.7133569 0.7468805 0.8076855
#>  [2,] 0.75801328 0.81422671 0.87182181 0.92431471 0.9626286 0.9510274 0.9579313
#>  [3,] 0.95910400 0.93590476 0.95986678 0.95596593 0.9441407 0.7379522 0.5865133
#>  [4,] 0.78293657 0.89690295 1.01754662 1.01560383 0.9353665 0.9236801 0.8641510
#>  [5,] 0.07400000 0.06250000 0.23810000 0.48240000 0.6372000 0.7531000 0.8959000
#>  [6,] 0.14617500 0.07507056 0.10983044 0.31286329 0.4411652 0.5740379 0.7757879
#>  [7,] 0.72527087 0.83183960 0.82211511 0.83513135 0.9490011 0.9563223 0.9854859
#>  [8,] 0.79531180 0.66563505 0.35061148 0.03410104 0.0277688 0.2464889 0.3899822
#>  [9,] 0.14964707 0.03327239 0.18904780 0.43732281 0.6008595 0.6923545 0.7599290
#> [10,] 0.09020000 0.15570000 0.32910000 0.52680000 0.6740000 0.7906000 0.8938000
#> [11,] 0.67483552 0.78599741 0.83684442 0.78017187 0.7980205 0.8366870 0.8933479
#> [12,] 0.88136247 0.80449607 0.80756246 0.70634592 0.4679876 0.3004740 0.1704644
#> [13,] 0.50347289 0.54387711 0.59632749 0.59411473 0.5773940 0.5697202 0.6151256
#> [14,] 0.32792931 0.50117684 0.73322939 0.83732553 0.9809221 0.8343047 0.6881386
#> [15,] 0.28752851 0.28176750 0.42341240 0.68569438 0.6772378 0.7657870 0.8605503
#> [16,] 0.35209416 0.56189701 0.73515845 0.83695133 0.8433833 0.8253706 0.6589907
#> [17,] 0.85908184 0.94809185 0.93969500 0.88059137 0.9081083 0.9208475 0.9637839
#> [18,] 0.39175811 0.49142433 0.63781713 0.73732866 0.7623657 0.7403648 0.7976128
#> [19,] 0.36674355 0.23577953 0.24350350 0.31696695 0.2845893 0.2967207 0.3653114
#> [20,] 0.86151919 0.79733076 0.68533573 0.43926249 0.3726354 0.5455820 0.2396306
#> [21,] 0.30009408 0.34835603 0.38290921 0.52188273 0.6894198 0.7930102 0.8351719
#> [22,] 0.05981807 0.11912807 0.07127833 0.20041361 0.2341382 0.1336100 0.2439511
#> [23,] 0.58411675 0.64442079 0.64157706 0.67924667 0.6973470 0.8982820 0.9031255
#> [24,] 0.44531545 0.43903847 0.31637057 0.33019768 0.4823786 0.4521825 0.5056503
#> [25,] 0.28255429 0.41436141 0.55770260 0.59994064 0.6398190 0.8123143 0.9572606
#>             V27       V28       V29         V3       V30       V31       V32
#>  [1,] 0.9164077 0.9204055 0.8029965 0.04964092 0.7467823 0.6992468 0.5849819
#>  [2,] 0.9452023 0.8827117 0.7564531 0.02641050 0.5990816 0.4227076 0.3370410
#>  [3,] 0.5381437 0.4679077 0.3991003 0.08644227 0.3430873 0.2638548 0.1753878
#>  [4,] 0.7824674 0.6851120 0.5680428 0.02410622 0.4974610 0.3200816 0.2065755
#>  [5,] 0.9941000 0.9957000 0.9328000 0.05470000 0.9344000 0.8854000 0.7690000
#>  [6,] 0.9244038 0.9459833 0.9604440 0.02584736 1.0321106 0.9526605 0.8311214
#>  [7,] 0.9983185 0.9906021 0.8936628 0.03199659 0.6352173 0.3826589 0.2864221
#>  [8,] 0.3945721 0.3591046 0.1929535 0.09986026 0.1670598 0.2900646 0.3093470
#>  [9,] 0.8940882 0.9372341 0.8745248 0.02494530 0.9144667 0.9604106 0.9261713
#> [10,] 0.9395000 0.9493000 0.9040000 0.00800000 0.9151000 0.8828000 0.8086000
#> [11,] 0.9508728 0.8710073 0.6953041 0.04266918 0.4976070 0.4337751 0.4848966
#> [12,] 0.2759402 0.2935351 0.2529862 0.01981502 0.1986762 0.1313817 0.2035644
#> [13,] 0.6534281 0.7654482 0.7631060 0.08237607 0.7897387 0.8027136 0.7640862
#> [14,] 0.6993409 0.5519195 0.4472881 0.10236818 0.5038761 0.4156571 0.1507777
#> [15,] 0.7516045 0.7921646 0.7561658 0.01987513 0.6500255 0.5134136 0.3883310
#> [16,] 0.6345615 0.6229128 0.3764804 0.05836501 0.4220495 0.5114376 0.4106654
#> [17,] 0.9361326 0.6853303 0.4091680 0.03437429 0.4003140 0.4851619 0.3776901
#> [18,] 0.7768965 0.7826586 0.8353663 0.02565719 0.7867852 0.6847755 0.6661420
#> [19,] 0.3733407 0.3681425 0.6073655 0.08386023 0.6863419 0.8522946 0.8230274
#> [20,] 0.5179972 0.8625278 0.6044809 0.11313877 0.8575121 0.8692287 0.5073541
#> [21,] 0.7750320 0.7341632 0.6546875 0.02499875 0.5075864 0.3931560 0.2949846
#> [22,] 0.4644256 0.6318246 0.6839488 0.01881691 0.6734739 0.8007017 0.6641868
#> [23,] 0.7981853 0.5861208 0.5072354 0.07928154 0.3253428 0.2799157 0.3989716
#> [24,] 0.6781844 0.8678744 0.9091805 0.03916734 0.8370129 0.7675983 0.6771125
#> [25,] 0.9533662 0.8822713 0.7772461 0.02174688 0.8176181 0.8592976 0.7393871
#>             V33        V34       V35        V36        V37        V38
#>  [1,] 0.4010596 0.31040384 0.1947760 0.07828665 0.23063675 0.21590484
#>  [2,] 0.2393727 0.11443524 0.1176259 0.13483476 0.19535893 0.23372067
#>  [3,] 0.3419186 0.37185409 0.3275459 0.25939288 0.33941489 0.46079346
#>  [4,] 0.1791859 0.23466900 0.3165684 0.34195359 0.30749308 0.21746710
#>  [5,] 0.6865000 0.63900000 0.6378000 0.66290000 0.59830000 0.45650000
#>  [6,] 0.7265828 0.68241700 0.7308464 0.77667555 0.76069527 0.61187798
#>  [7,] 0.2129186 0.18915658 0.1830157 0.09874934 0.06151084 0.09168643
#>  [8,] 0.3338850 0.33366670 0.4213099 0.07324586 0.12486615 0.19882982
#>  [9,] 0.7728746 0.65546632 0.6423686 0.67427823 0.69983410 0.59295764
#> [10,] 0.7180000 0.67200000 0.6447000 0.68790000 0.62410000 0.49360000
#> [11,] 0.4271196 0.29977514 0.1037399 0.29511932 0.33765245 0.35420517
#> [12,] 0.2132928 0.16807370 0.2561063 0.10732460 0.22870786 0.25238222
#> [13,] 0.5931063 0.49619112 0.4410827 0.25635978 0.22315594 0.34518772
#> [14,] 0.1675074 0.06739692 0.3987622 0.68262055 0.67902170 0.31167091
#> [15,] 0.5320945 0.75048160 0.7865443 0.61966444 0.29565902 0.20906641
#> [16,] 0.3002670 0.26078805 0.5512866 0.69177181 0.69080211 0.49304188
#> [17,] 0.3014481 0.28496374 0.2073146 0.09708772 0.25450181 0.20470332
#> [18,] 0.6999043 0.70679087 0.6647554 0.59894482 0.50368906 0.42307857
#> [19,] 0.6032205 0.55451097 0.6038534 0.65662187 0.76836983 0.96227316
#> [20,] 0.1633940 0.24784394 0.4010133 0.28057080 0.59902187 0.66580291
#> [21,] 0.2482595 0.27243490 0.2893933 0.26163170 0.15890460 0.14393937
#> [22,] 0.7009257 0.87421758 0.9826012 0.96280768 0.75390606 0.58677585
#> [23,] 0.4810172 0.44414259 0.5018519 0.59451570 0.59946638 0.28812638
#> [24,] 0.6506663 0.53847190 0.4931186 0.54975025 0.53332273 0.36142684
#> [25,] 0.6714213 0.64928512 0.6984623 0.70531269 0.66297171 0.58179926
#>              V39         V4        V40        V41        V42        V43
#>  [1,] 0.16257178 0.05962823 0.24599075 0.31713469 0.21153625 0.19952538
#>  [2,] 0.23074165 0.04963967 0.26004611 0.25825712 0.24144571 0.20618407
#>  [3,] 0.45564370 0.04959669 0.44584307 0.37746168 0.40668930 0.36096193
#>  [4,] 0.29152297 0.09280214 0.36638922 0.43554207 0.45717305 0.31096399
#>  [5,] 0.31290000 0.06810000 0.41580000 0.43250000 0.40310000 0.42010000
#>  [6,] 0.50018405 0.07071063 0.51522415 0.45129287 0.39642115 0.47673800
#>  [7,] 0.07741708 0.03912721 0.06903557 0.05388477 0.10770330 0.19654678
#>  [8,] 0.19658444 0.13925627 0.12294235 0.05165790 0.09437034 0.10936493
#>  [9,] 0.56545242 0.02287283 0.61942178 0.58177321 0.46482570 0.49192909
#> [10,] 0.41440000 0.05540000 0.42400000 0.45460000 0.43920000 0.43230000
#> [11,] 0.45860174 0.03569590 0.38389634 0.20083237 0.29405198 0.39344797
#> [12,] 0.19395958 0.05821117 0.09257849 0.07008336 0.27116861 0.22687885
#> [13,] 0.36260182 0.08056316 0.26879409 0.55376583 0.65872628 0.49148528
#> [14,] 0.09221061 0.08627741 0.32936105 0.42728032 0.38021911 0.23307258
#> [15,] 0.30203237 0.04185605 0.34805996 0.25084754 0.31526296 0.36783356
#> [16,] 0.29718855 0.04936381 0.37589526 0.37801457 0.33499176 0.27917352
#> [17,] 0.12437965 0.03216701 0.16661892 0.08427018 0.01732845 0.06052893
#> [18,] 0.33235458 0.03444097 0.32488359 0.26263281 0.16348928 0.16977681
#> [19,] 0.94653340 0.01819280 0.83300485 0.76125353 0.65938987 0.52382164
#> [20,] 0.53004159 0.10924624 0.47162305 0.47454710 0.26089636 0.21466940
#> [21,] 0.17303331 0.02523099 0.17240046 0.19324462 0.16833219 0.15775462
#> [22,] 0.56326054 0.01231811 0.59285009 0.51314366 0.46769110 0.34295470
#> [23,] 0.18362734 0.05655374 0.40452519 0.40426446 0.51763556 0.44093737
#> [24,] 0.21269872 0.03690604 0.12452552 0.12612715 0.17526820 0.17710293
#> [25,] 0.37267931 0.02801472 0.22126437 0.22756974 0.21115209 0.24013237
#>              V44        V45         V46        V47        V48        V49
#>  [1,] 0.17293016 0.16568249 0.125443894 0.11706553 0.11912944 0.07409965
#>  [2,] 0.17709428 0.14395001 0.116906196 0.11595148 0.08711753 0.05057791
#>  [3,] 0.25138827 0.15421548 0.109370934 0.06601791 0.02575999 0.03654428
#>  [4,] 0.19869150 0.13499765 0.197404404 0.21242595 0.12531043 0.06895064
#>  [5,] 0.45570000 0.39550000 0.296600000 0.20950000 0.15580000 0.08840000
#>  [6,] 0.54576569 0.45871489 0.265522601 0.14680167 0.14494136 0.09723159
#>  [7,] 0.22694734 0.22490474 0.234705356 0.18408167 0.11483656 0.05933422
#>  [8,] 0.08193254 0.06793422 0.042737181 0.06289373 0.07294486 0.04087539
#>  [9,] 0.57826508 0.53282719 0.432921515 0.34241822 0.24572375 0.12149249
#> [10,] 0.49210000 0.47100000 0.319600000 0.22410000 0.18060000 0.09900000
#> [11,] 0.43815815 0.41272697 0.197797396 0.10823254 0.11269689 0.06696035
#> [12,] 0.03757280 0.19597511 0.139952971 0.06056786 0.07302300 0.05685143
#> [13,] 0.41250901 0.54787458 0.533852079 0.32517022 0.19263156 0.11008845
#> [14,] 0.18601241 0.21213998 0.172308747 0.17387874 0.16372307 0.11277945
#> [15,] 0.24861623 0.08786496 0.035125776 0.05541887 0.05540833 0.02445267
#> [16,] 0.22451717 0.22147461 0.189236382 0.16820352 0.14292669 0.08959093
#> [17,] 0.13510910 0.12293383 0.048766181 0.06896162 0.05569280 0.02015219
#> [18,] 0.18128203 0.14745627 0.098592905 0.04704068 0.02928781 0.01856658
#> [19,] 0.42241459 0.21941547 0.184082215 0.17884756 0.14398585 0.06739886
#> [20,] 0.21969976 0.20239532 0.006364307 0.13391530 0.07513469 0.01221531
#> [21,] 0.12282696 0.09408190 0.088714923 0.05879139 0.03991200 0.02847938
#> [22,] 0.22371390 0.14205837 0.077043185 0.08732182 0.07135611 0.04588969
#> [23,] 0.30644908 0.15591469 0.074925274 0.03922956 0.04913888 0.02537456
#> [24,] 0.16983349 0.13614546 0.097010580 0.06287789 0.07819638 0.04431326
#> [25,] 0.19175695 0.16316022 0.151713236 0.08421358 0.04258347 0.02836660
#>               V5         V50         V51         V52         V53          V54
#>  [1,] 0.05359533 0.016920414 0.019792014 0.017618245 0.008049139 0.0110494366
#>  [2,] 0.09098756 0.019015321 0.015517093 0.014241693 0.012592021 0.0111358210
#>  [3,] 0.11304801 0.012399135 0.008599321 0.009741168 0.017422326 0.0211201513
#>  [4,] 0.12067804 0.031954778 0.023239369 0.017071551 0.021038307 0.0168914401
#>  [5,] 0.07840000 0.026500000 0.012100000 0.009100000 0.006200000 0.0019000000
#>  [6,] 0.10624278 0.027252553 0.010758405 0.017685076 0.011659226 0.0112348844
#>  [7,] 0.09153361 0.015457187 0.025126876 0.015841403 0.004712130 0.0009908597
#>  [8,] 0.06953841 0.020113316 0.031035351 0.026652011 0.024474785 0.0104568694
#>  [9,] 0.06330683 0.043844019 0.026595622 0.015243952 0.004470664 0.0095997619
#> [10,] 0.08830000 0.025100000 0.012900000 0.009500000 0.012600000 0.0069000000
#> [11,] 0.05654978 0.015372496 0.023141995 0.012501376 0.006101742 0.0100909944
#> [12,] 0.17205636 0.043310098 0.026664459 0.028124200 0.022621550 0.0176785216
#> [13,] 0.11805806 0.046621653 0.041161156 0.031097548 0.020441243 0.0165253780
#> [14,] 0.05423394 0.025960752 0.019415692 0.009704253 0.005099039 0.0126420664
#> [15,] 0.06025149 0.011577235 0.025040962 0.009561205 0.009517570 0.0120280663
#> [16,] 0.07948258 0.030326186 0.012800980 0.019293283 0.013481934 0.0116257497
#> [17,] 0.05960358 0.012702338 0.014729954 0.008733024 0.012498045 0.0075689517
#> [18,] 0.06166563 0.011320805 0.006588546 0.009360624 0.009366049 0.0081253968
#> [19,] 0.03958081 0.025176934 0.015213543 0.012411524 0.016291950 0.0032501345
#> [20,] 0.09826931 0.010203808 0.002765333 0.023400019 0.016815405 0.0089653333
#> [21,] 0.05171193 0.013643156 0.012135747 0.015024687 0.009046870 0.0069217986
#> [22,] 0.04602610 0.016166775 0.007132787 0.005060697 0.014072462 0.0099093020
#> [23,] 0.03627605 0.006929935 0.001326804 0.008844867 0.013063076 0.0169234310
#> [24,] 0.02382059 0.011838520 0.009233474 0.008779929 0.007067690 0.0113872452
#> [25,] 0.03338948 0.023375419 0.004714722 0.008561042 0.001858889 0.0054234393
#>               V55         V56         V57          V58         V59         V6
#>  [1,] 0.007868411 0.009728584 0.007279596 0.0082140369 0.005487304 0.10791865
#>  [2,] 0.004972584 0.010243997 0.006654030 0.0059278924 0.006496070 0.13388178
#>  [3,] 0.010182745 0.016795058 0.013511087 0.0141331298 0.015279439 0.17176384
#>  [4,] 0.003136279 0.008852374 0.010526169 0.0090384305 0.010517244 0.05728277
#>  [5,] 0.004500000 0.007900000 0.003100000 0.0063000000 0.004800000 0.12500000
#>  [6,] 0.001359848 0.007947572 0.007167442 0.0055088486 0.001265252 0.10615147
#>  [7,] 0.002850946 0.005400891 0.002630816 0.0046271218 0.001713998 0.08505527
#>  [8,] 0.017288024 0.018319733 0.005725008 0.0318119645 0.016083638 0.08700839
#>  [9,] 0.009232874 0.008672416 0.008396257 0.0017671334 0.004117420 0.07650936
#> [10,] 0.003900000 0.006800000 0.006000000 0.0045000000 0.000200000 0.12780000
#> [11,] 0.010410688 0.005842886 0.006242536 0.0061481141 0.005633058 0.11590374
#> [12,] 0.023672958 0.004798016 0.000739140 0.0128972893 0.013832860 0.21136161
#> [13,] 0.016222301 0.013516894 0.014215578 0.0161323403 0.012981407 0.11522749
#> [14,] 0.007710865 0.005475756 0.009761362 0.0079590628 0.010326218 0.07970220
#> [15,] 0.005466896 0.011600461 0.009980554 0.0041613720 0.008111911 0.08976696
#> [16,] 0.006586766 0.005682658 0.017113467 0.0089838270 0.005801651 0.11944198
#> [17,] 0.005392769 0.008964664 0.005540645 0.0003068092 0.003986343 0.08279110
#> [18,] 0.006938040 0.008742647 0.007154490 0.0033169680 0.003438983 0.09048118
#> [19,] 0.001416901 0.004783743 0.001035491 0.0143963079 0.010234452 0.06196523
#> [20,] 0.017630734 0.025215463 0.032492393 0.0165615540 0.009550005 0.23244273
#> [21,] 0.007467151 0.005523969 0.005696961 0.0056554837 0.004487881 0.08268155
#> [22,] 0.008727696 0.005769969 0.008269680 0.0076609791 0.007818551 0.08229898
#> [23,] 0.012819543 0.016670827 0.012514243 0.0067770745 0.008583124 0.07445505
#> [24,] 0.008797748 0.006218094 0.009510463 0.0072538115 0.009533260 0.08236184
#> [25,] 0.007819931 0.003390034 0.011597396 0.0109722558 0.007000000 0.06713795
#>               V60         V7         V8          V9
#>  [1,] 0.007779079 0.11341478 0.10828264  0.18793579
#>  [2,] 0.006127058 0.15519484 0.16367248  0.21442632
#>  [3,] 0.009990115 0.16056839 0.10573748  0.09937109
#>  [4,] 0.008666102 0.11815017 0.10679896  0.05716117
#>  [5,] 0.005000000 0.12960000 0.17290000  0.27940000
#>  [6,] 0.003116767 0.15733994 0.21530887  0.33033759
#>  [7,] 0.003829236 0.09947237 0.13253709  0.20495212
#>  [8,] 0.016502735 0.15073346 0.09932042  0.25008500
#>  [9,] 0.002562258 0.11333852 0.11894987  0.23341714
#> [10,] 0.002900000 0.16740000 0.13730000  0.29220000
#> [11,] 0.004638182 0.10685364 0.16714346  0.28160218
#> [12,] 0.011028985 0.16966580 0.09173623  0.10117784
#> [13,] 0.009128226 0.10785717 0.13595227  0.20346590
#> [14,] 0.006470443 0.14693255 0.20172996  0.22768396
#> [15,] 0.006665138 0.10651920 0.10292191  0.08626400
#> [16,] 0.005837038 0.18724460 0.24556980  0.27540609
#> [17,] 0.006296445 0.14040829 0.09588502  0.10792255
#> [18,] 0.004415842 0.10898582 0.08005406  0.10973546
#> [19,] 0.007447382 0.04172629 0.06249073  0.11267712
#> [20,] 0.007549976 0.23967275 0.37394970  0.56007695
#> [21,] 0.003734061 0.08954779 0.09981274  0.10875535
#> [22,] 0.003490706 0.05620878 0.01307907 -0.04614976
#> [23,] 0.012805739 0.07854546 0.11185591  0.01981974
#> [24,] 0.005705507 0.07697377 0.06510748  0.12442652
#> [25,] 0.008585278 0.10992740 0.12056375  0.07674281
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841