Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1        V10         V11        V12        V13        V14
#>  [1,]  0.030581676 0.40489562  0.48611272 0.51147762 0.48915066 0.38940207
#>  [2,]  0.017803959 0.09389692  0.14921986 0.13295585 0.07866773 0.13960532
#>  [3,]  0.167383506 0.16676798  0.46025372 0.56236341 0.71364633 0.67571365
#>  [4,]  0.028359274 0.13708278  0.22848997 0.34838362 0.41830323 0.46420725
#>  [5,]  0.035881827 0.24166348  0.28192445 0.30544304 0.27738374 0.22097662
#>  [6,]  0.022500611 0.21734613  0.20535095 0.14026791 0.16957697 0.29527902
#>  [7,]  0.045579575 0.40222844  0.32410172 0.23853804 0.21937724 0.23783065
#>  [8,]  0.023301042 0.20886120  0.22351746 0.23060904 0.31114481 0.38475807
#>  [9,]  0.018038793 0.31353070  0.48877508 0.53716649 0.45030962 0.34254112
#> [10,]  0.029262391 0.37601946  0.38691971 0.19944249 0.47119487 0.92349542
#> [11,]  0.020469417 0.25221526  0.29551718 0.34848247 0.29812258 0.21426964
#> [12,]  0.019200896 0.12164869  0.22127686 0.27444236 0.23996606 0.15300254
#> [13,]  0.080175275 0.05494375  0.15511486 0.19099271 0.16159157 0.22997855
#> [14,]  0.015718513 0.05530929  0.18437897 0.26297950 0.17165374 0.16329314
#> [15,]  0.011379400 0.07943606  0.06940929 0.09492183 0.06896123 0.12710105
#> [16,] -0.002962059 0.01892817 -0.02719213 0.16022502 0.13396415 0.07004364
#> [17,]  0.014006012 0.15173949  0.11900974 0.06230534 0.04967730 0.09752193
#> [18,]  0.018313208 0.12408387  0.13495234 0.10346315 0.12104011 0.10608635
#> [19,]  0.012570814 0.05236142  0.18317787 0.37169165 0.27223291 0.22355307
#> [20,]  0.025684218 0.11804155  0.21913088 0.19909428 0.27754880 0.32356483
#> [21,]  0.023140321 0.10097177  0.08541412 0.07557922 0.18080887 0.32143808
#> [22,]  0.029300000 0.08600000  0.04140000 0.04720000 0.08350000 0.09380000
#> [23,]  0.030394102 0.28497107  0.23126670 0.22965051 0.24394422 0.11025139
#> [24,]  0.010195115 0.08571880  0.09816225 0.09606274 0.14335476 0.17283750
#> [25,]  0.018249787 0.14943355  0.16001917 0.19523446 0.27151836 0.28111822
#>              V15        V16        V17       V18        V19            V2
#>  [1,] 0.26360426 0.25856097 0.29037986 0.3297456 0.47622108  3.161717e-02
#>  [2,] 0.19650627 0.38884951 0.46326105 0.6069906 0.70566368 -3.834904e-03
#>  [3,] 0.64033653 0.76878965 0.97319303 0.9302863 0.96728190  1.519688e-01
#>  [4,] 0.51067390 0.54037095 0.64410464 0.8098556 0.95646932  5.618767e-02
#>  [5,] 0.17681810 0.18082294 0.19584157 0.2463804 0.29521910  4.076442e-02
#>  [6,] 0.43440858 0.57810054 0.72481846 0.6711406 0.75773639  2.185550e-02
#>  [7,] 0.25577114 0.29818564 0.23102515 0.2294371 0.26092580  4.906117e-02
#>  [8,] 0.42510706 0.50371246 0.59627252 0.6756950 0.74712976  3.889771e-02
#>  [9,] 0.27547766 0.38503512 0.44544564 0.3955555 0.40555974  1.501431e-02
#> [10,] 0.91118380 0.85005853 0.72544015 0.8133662 0.91166451  3.652695e-02
#> [11,] 0.17036375 0.23233232 0.25202759 0.3648060 0.56271819  2.193145e-02
#> [12,] 0.07826217 0.17780665 0.29794022 0.3937509 0.47996801  3.448815e-02
#> [13,] 0.29993790 0.52915407 0.66390856 0.6634119 0.77382313  1.092464e-01
#> [14,] 0.25590811 0.37660339 0.38719270 0.2795678 0.36662763  5.741525e-02
#> [15,] 0.05071279 0.13754689 0.13120280 0.3275289 0.35049369  1.270493e-02
#> [16,] 0.09453684 0.07291115 0.06764489 0.2079488 0.03722549 -2.388079e-03
#> [17,] 0.08939651 0.15270101 0.26034215 0.2851593 0.36036689  2.726732e-03
#> [18,] 0.14259293 0.23778390 0.26966646 0.2946307 0.34086130  1.510161e-02
#> [19,] 0.09005533 0.08616662 0.16960602 0.3983173 0.28276836  5.112962e-02
#> [20,] 0.43960961 0.54905252 0.42023758 0.4396371 0.62389189  4.214254e-02
#> [21,] 0.51951286 0.70642234 0.77808350 0.7812212 0.83242102  2.140653e-02
#> [22,] 0.14660000 0.08090000 0.11790000 0.2179000 0.33260000  6.440000e-02
#> [23,] 0.33937312 0.31990136 0.22542353 0.2225879 0.23494021 -2.730042e-05
#> [24,] 0.16038489 0.18134122 0.19072420 0.2121230 0.24535965  1.157460e-02
#> [25,] 0.32968452 0.37909768 0.38681213 0.3306202 0.23720275  2.203296e-02
#>              V20       V21       V22        V23       V24       V25       V26
#>  [1,] 0.68357927 0.7848820 0.8203316 0.76195202 0.7911641 0.8021066 0.8859642
#>  [2,] 0.76195240 0.8605783 0.9640841 1.00635071 0.9964182 1.0102314 0.8933779
#>  [3,] 0.84164614 0.9461112 0.8127931 0.62665035 0.5688449 0.4583007 0.4383903
#>  [4,] 1.00395108 1.0330686 1.0582035 0.93601689 0.8639541 0.6755797 0.5596346
#>  [5,] 0.31996158 0.4137266 0.5363353 0.66211842 0.7405424 0.7901355 0.8202894
#>  [6,] 0.90047606 1.0028809 0.9876120 0.78449038 0.5378287 0.3647974 0.4537212
#>  [7,] 0.39634791 0.4091630 0.2685634 0.30358142 0.4159559 0.4973137 0.6329639
#>  [8,] 0.80466979 0.8954619 0.9226502 0.94138445 0.9388729 0.9112398 0.9173428
#>  [9,] 0.49050904 0.7248880 0.9483605 1.00929052 0.9919723 1.0070358 0.9224569
#> [10,] 0.87936794 0.8984657 0.9421967 0.98569424 0.8874454 0.7631097 0.7832034
#> [11,] 0.66969435 0.7857204 0.7800153 0.81872682 0.9367840 0.9327064 0.9422294
#> [12,] 0.59059844 0.6687584 0.7650239 0.80406500 0.7510399 0.7912611 0.9132977
#> [13,] 0.69923259 0.5913610 0.5539630 0.62943300 0.5911519 0.3987961 0.3397675
#> [14,] 0.32634102 0.4830326 0.4214549 0.47223887 0.6542980 0.5785269 0.6199094
#> [15,] 0.29960509 0.1921095 0.2141219 0.27931028 0.2957952 0.1348838 0.2747909
#> [16,] 0.03962536 0.2026126 0.1209103 0.04965674 0.1261282 0.0159530 0.1366528
#> [17,] 0.53011257 0.6389576 0.8100342 0.95263350 0.8836656 0.8633660 0.9444800
#> [18,] 0.39148426 0.4874279 0.5666019 0.65323905 0.6666565 0.6794347 0.7699750
#> [19,] 0.58764691 0.7134245 0.7455230 0.77758071 0.8310683 1.0216473 0.9619533
#> [20,] 0.64046378 0.5026568 0.5105999 0.73011701 0.8577985 0.7962812 0.8504120
#> [21,] 0.85855609 0.9690268 0.9608297 0.93685584 0.9367969 0.8922826 0.9270398
#> [22,] 0.32580000 0.2111000 0.2302000 0.33610000 0.4259000 0.4609000 0.2606000
#> [23,] 0.39215309 0.2522633 0.6900766 0.66154738 0.3391270 0.2341487 0.6000413
#> [24,] 0.24310327 0.2648966 0.3634736 0.44310743 0.5625802 0.6111424 0.6693727
#> [25,] 0.32178474 0.4267266 0.4930921 0.52974640 0.6823926 0.8221445 0.7640563
#>             V27       V28       V29           V3       V30        V31       V32
#>  [1,] 0.9552574 0.8821737 0.6988645 0.0409969587 0.5091453 0.43967982 0.5127420
#>  [2,] 0.6899850 0.5067993 0.4046809 0.0177857608 0.4599451 0.38660175 0.3207874
#>  [3,] 0.3531228 0.5036791 0.3375811 0.1700473420 0.1950433 0.05691605 0.1196024
#>  [4,] 0.5353924 0.4073370 0.3563574 0.0637314408 0.3225791 0.25269550 0.1388100
#>  [5,] 0.8638405 0.8987036 0.8521269 0.0462650453 0.6887354 0.63642777 0.5281941
#>  [6,] 0.4902959 0.2725943 0.2456146 0.0019046050 0.4702773 0.23145918 0.3189868
#>  [7,] 0.6679406 0.7755648 0.7719303 0.0391540688 0.8503524 0.69855530 0.6304010
#>  [8,] 0.9270902 0.8945638 0.7698872 0.0367317586 0.5983011 0.40971697 0.2830790
#>  [9,] 0.8262913 0.7005081 0.4616241 0.0259437855 0.3356424 0.24979876 0.4010933
#> [10,] 0.6166256 0.4446871 0.4992499 0.0210481065 0.3599291 0.28952340 0.1177340
#> [11,] 0.9682501 0.9843064 0.8720628 0.0352522536 0.6389086 0.40726740 0.2786254
#> [12,] 1.0016119 0.9700130 0.8073747 0.0481387876 0.6340652 0.30092563 0.2455068
#> [13,] 0.4446869 0.4553984 0.5763791 0.1328545045 0.4242826 0.16454456 0.2174362
#> [14,] 0.7741075 0.7232948 0.6031481 0.0698252790 0.7595772 0.59897704 0.5003520
#> [15,] 0.4577073 0.6007618 0.7136430 0.0233477425 0.6789474 0.57504940 0.3308237
#> [16,] 0.3525935 0.4877770 0.5303786 0.0007845119 0.5330721 0.66855770 0.6445231
#> [17,] 0.9595115 0.9085033 0.8246870 0.0036770685 0.7104388 0.60662438 0.5455685
#> [18,] 0.7927611 0.7715002 0.7217362 0.0231362725 0.6483972 0.59288910 0.5468000
#> [19,] 0.7809232 0.6034770 0.5348000 0.0694271444 0.5019136 0.50338701 0.4481728
#> [20,] 0.8542811 0.8094283 0.7596442 0.0548341945 0.7322047 0.52731969 0.2151556
#> [21,] 0.8749702 0.6850589 0.4473635 0.0327272822 0.4249838 0.43560226 0.3428364
#> [22,] 0.0874000 0.2862000 0.5606000 0.0390000000 0.8344000 0.80960000 0.7250000
#> [23,] 0.9160274 0.7187902 0.7483502 0.0054029715 0.9453549 0.63873952 0.3451496
#> [24,] 0.7150201 0.8431040 0.9434624 0.0187820182 0.9102405 0.80210183 0.7112115
#> [25,] 0.7918739 0.8822911 0.7663049 0.0079542658 0.5713790 0.52496117 0.4206110
#>              V33       V34       V35         V36        V37        V38
#>  [1,] 0.47921875 0.3437747 0.1334381  0.30112990 0.32556283 0.36024090
#>  [2,] 0.32825605 0.4114875 0.4011995  0.34029846 0.23535105 0.17027486
#>  [3,] 0.28730681 0.2753411 0.3563202 -0.03679627 0.04437276 0.24834029
#>  [4,] 0.32208209 0.3129962 0.2427628  0.28446660 0.43056127 0.41184700
#>  [5,] 0.46127988 0.3342612 0.1985057  0.14244745 0.19970232 0.15704263
#>  [6,] 0.43465437 0.2639117 0.1858513  0.33595501 0.35987323 0.31740407
#>  [7,] 0.63503155 0.6693319 0.7105345  0.71257253 0.66647464 0.64097390
#>  [8,] 0.20782569 0.1444240 0.1800221  0.16772233 0.13869693 0.19396033
#>  [9,] 0.42559239 0.3314003 0.1361883  0.18912351 0.31920671 0.33370404
#> [10,] 0.04708953 0.1431093 0.1846929  0.19829641 0.12961210 0.16018564
#> [11,] 0.21490753 0.1604751 0.1305385  0.05022939 0.06171378 0.09147921
#> [12,] 0.27202762 0.4772035 0.7264832  0.84501425 0.76138858 0.62068107
#> [13,] 0.23923094 0.2892562 0.4161670  0.42402472 0.41139636 0.45125259
#> [14,] 0.35161524 0.3891971 0.2669936  0.10242901 0.27012748 0.37241468
#> [15,] 0.50741550 0.4437761 0.2463823  0.30509961 0.47817900 0.62163226
#> [16,] 0.68894803 0.9124010 1.0258276  0.96351053 0.85173883 0.67532757
#> [17,] 0.52825168 0.5754694 0.6072007  0.54712998 0.49852762 0.36239536
#> [18,] 0.51388721 0.5929154 0.6707065  0.68797548 0.56920557 0.45098973
#> [19,] 0.53655882 0.5021659 0.5689074  0.61303076 0.53702738 0.27672140
#> [20,] 0.30123882 0.1395984 0.2836544  0.58501677 0.51867774 0.14604234
#> [21,] 0.30328077 0.2865751 0.1881872  0.07974041 0.18293100 0.14659685
#> [22,] 0.80480000 0.9435000 1.0000000  0.89600000 0.55160000 0.30370000
#> [23,] 0.69892905 0.9154417 0.7949652  0.75651924 0.56469139 0.41681968
#> [24,] 0.68382707 0.5989268 0.5393984  0.48017744 0.38429625 0.27646347
#> [25,] 0.25591186 0.1968200 0.2371166  0.28162677 0.21728416 0.22372256
#>               V39          V4       V40         V41        V42        V43
#>  [1,] 0.462367733 0.033729064 0.3874167 0.212906678 0.35929795 0.41632751
#>  [2,] 0.269454205 0.065945639 0.3474264 0.410149015 0.43419944 0.27762734
#>  [3,] 0.261666904 0.184724153 0.1194630 0.003394118 0.26794158 0.40920759
#>  [4,] 0.351211214 0.021574351 0.3794017 0.292935085 0.35184098 0.27846561
#>  [5,] 0.194796839 0.057501799 0.2300396 0.276681726 0.27651449 0.21425012
#>  [6,] 0.204881635 0.015389332 0.1177833 0.064090435 0.22794660 0.18886455
#>  [7,] 0.557852356 0.069306288 0.5003994 0.443487946 0.39043805 0.39493903
#>  [8,] 0.221216776 0.047743248 0.2433823 0.212678787 0.19234009 0.16109924
#>  [9,] 0.572256321 0.028195350 0.5571116 0.325490828 0.03703390 0.31831394
#> [10,] 0.315766399 0.061201032 0.3300340 0.214463481 0.12178052 0.26256897
#> [11,] 0.095948878 0.042064194 0.1010447 0.064626895 0.11103985 0.19672835
#> [12,] 0.471876244 0.061531267 0.3511014 0.380191559 0.34893375 0.38802939
#> [13,] 0.377840806 0.204543643 0.3965254 0.341733549 0.37326193 0.42624680
#> [14,] 0.214179028 0.065822835 0.3766909 0.426337636 0.19459128 0.25290537
#> [15,] 0.693067396 0.025643602 0.5768481 0.322084199 0.26683266 0.18088071
#> [16,] 0.657797998 0.042852247 0.6522964 0.527919709 0.44223088 0.30114770
#> [17,] 0.182244604 0.025417821 0.1172977 0.134797289 0.04846358 0.06721065
#> [18,] 0.383010809 0.029269764 0.3299426 0.232392695 0.17794075 0.19072109
#> [19,] 0.001403728 0.058559777 0.2526146 0.317151550 0.52310240 0.46274421
#> [20,] 0.148419917 0.068221033 0.1484736 0.357026597 0.37915503 0.26000258
#> [21,] 0.090417735 0.042714369 0.1309855 0.098984250 0.05135343 0.06821370
#> [22,] 0.233800000 0.017300000 0.2382000 0.331800000 0.38210000 0.15750000
#> [23,] 0.016190812 0.031252467 0.6551802 0.781530656 0.62477274 0.33610882
#> [24,] 0.213636157 0.025723759 0.1838144 0.152256786 0.14011606 0.20326125
#> [25,] 0.206477408 0.006583824 0.1650461 0.180456483 0.20625654 0.19329237
#>              V44        V45        V46        V47        V48         V49
#>  [1,] 0.46223869 0.43161403 0.21513587 0.12843556 0.12922683 0.070167543
#>  [2,] 0.16446390 0.10158675 0.19138395 0.20734912 0.12740569 0.066644966
#>  [3,] 0.15316377 0.06869785 0.10816920 0.06196796 0.06204171 0.057587579
#>  [4,] 0.14126017 0.07211888 0.09928447 0.05350562 0.02790906 0.032025842
#>  [5,] 0.13238343 0.11850915 0.13789433 0.15522701 0.13400267 0.066733702
#>  [6,] 0.11774830 0.08461628 0.02890195 0.05731554 0.07237890 0.043741553
#>  [7,] 0.44725678 0.49330872 0.38092196 0.23867597 0.20351409 0.110064566
#>  [8,] 0.16771780 0.16332903 0.11447975 0.08815318 0.05936393 0.042108051
#>  [9,] 0.34650098 0.42288304 0.37231919 0.21093166 0.08715178 0.060972141
#> [10,] 0.25228443 0.13437939 0.11228638 0.13504600 0.11891453 0.094693803
#> [11,] 0.22419421 0.20672050 0.21642620 0.17634601 0.10994823 0.052087084
#> [12,] 0.33474572 0.32531333 0.23841976 0.09745073 0.03131400 0.042495070
#> [13,] 0.34865429 0.21702006 0.19098188 0.20058378 0.12470841 0.059905147
#> [14,] 0.29423024 0.22560981 0.13163445 0.05718897 0.09896118 0.079364074
#> [15,] 0.11146238 0.09315887 0.07573111 0.10908759 0.08550602 0.028231355
#> [16,] 0.16111730 0.10203661 0.06312679 0.06963315 0.06688393 0.024765378
#> [17,] 0.03326116 0.01836272 0.02568349 0.02344960 0.03074160 0.013258435
#> [18,] 0.21202277 0.13009931 0.08958964 0.04949488 0.04136312 0.019725664
#> [19,] 0.26611387 0.17153352 0.12409529 0.06968036 0.06912054 0.021055736
#> [20,] 0.12453671 0.15746511 0.17151598 0.09539007 0.08831027 0.034399964
#> [21,] 0.12089925 0.10805027 0.04610582 0.06749109 0.05111008 0.027117961
#> [22,] 0.22280000 0.15820000 0.14330000 0.16340000 0.11330000 0.056700000
#> [23,] 0.14991910 0.21599286 0.34461910 0.22430559 0.02977226 0.001020204
#> [24,] 0.21323468 0.16940424 0.10845072 0.05956025 0.05032060 0.027777389
#> [25,] 0.15188615 0.11202230 0.09228441 0.06742802 0.04467593 0.023172579
#>               V5         V50         V51         V52          V53         V54
#>  [1,] 0.07319303 0.017465902 0.022274582 0.013526594 0.0074285416 0.010555487
#>  [2,] 0.08968244 0.028668869 0.023320183 0.016102959 0.0196029024 0.015709593
#>  [3,] 0.21122648 0.039543107 0.028229037 0.019744394 0.0263170286 0.042222579
#>  [4,] 0.09447072 0.019308339 0.014673739 0.019564935 0.0182593855 0.023015638
#>  [5,] 0.07220722 0.014709580 0.018071613 0.016018569 0.0080414579 0.007570159
#>  [6,] 0.03103246 0.022812618 0.015575618 0.011656414 0.0008822779 0.010598358
#>  [7,] 0.07822081 0.036448474 0.029589020 0.021506463 0.0130610318 0.016090411
#>  [8,] 0.13685177 0.016588548 0.019234904 0.011836793 0.0160749345 0.011158520
#>  [9,] 0.04246694 0.027111836 0.012571626 0.003623558 0.0134534661 0.013948408
#> [10,] 0.08142056 0.029034669 0.032554627 0.007566799 0.0146585091 0.011297343
#> [11,] 0.07302508 0.015015253 0.018550008 0.012884128 0.0058330508 0.004592909
#> [12,] 0.04334465 0.017528643 0.004342369 0.013162526 0.0096239194 0.003269596
#> [13,] 0.19684380 0.021692978 0.010715622 0.014785901 0.0206248556 0.014948254
#> [14,] 0.03215734 0.016981826 0.020500357 0.024714301 0.0155313378 0.021252386
#> [15,] 0.03999873 0.012259098 0.008748504 0.017988515 0.0217501967 0.015801204
#> [16,] 0.01998523 0.021881060 0.023916329 0.002227252 0.0034951752 0.005076560
#> [17,] 0.05276459 0.007597507 0.002442708 0.004166798 0.0088833040 0.006736228
#> [18,] 0.06272322 0.011508145 0.010031079 0.006634125 0.0057591670 0.009103451
#> [19,] 0.12404997 0.003352328 0.005242738 0.003362890 0.0116289893 0.009155700
#> [20,] 0.04307169 0.021091370 0.010150418 0.014691648 0.0124737277 0.010873613
#> [21,] 0.06401012 0.014919699 0.012884261 0.009238711 0.0110908319 0.007241071
#> [22,] 0.04760000 0.013300000 0.017000000 0.003500000 0.0052000000 0.008300000
#> [23,] 0.06613134 0.051921589 0.041235506 0.036594308 0.0152081088 0.012765084
#> [24,] 0.02701527 0.012735699 0.010513848 0.008520862 0.0093358962 0.007337771
#> [25,] 0.02982706 0.011943715 0.008777731 0.004443513 0.0035919303 0.007130161
#>               V55         V56          V57          V58         V59         V6
#>  [1,] 0.008386742 0.005603692 0.0048734063  0.004584513 0.006742231 0.11998084
#>  [2,] 0.003749775 0.007347468 0.0090754231  0.006859003 0.007185628 0.06569287
#>  [3,] 0.012605926 0.007682835 0.0003494488  0.013386662 0.019823415 0.13067908
#>  [4,] 0.008076921 0.011029857 0.0116970436  0.012488866 0.011176678 0.15132755
#>  [5,] 0.005900653 0.007322147 0.0064196420  0.007583033 0.006303757 0.12109937
#>  [6,] 0.003910029 0.009773358 0.0042211428  0.007599595 0.001218381 0.01928827
#>  [7,] 0.011333575 0.011291152 0.0113687449  0.008541942 0.007929561 0.11095627
#>  [8,] 0.007985356 0.007034886 0.0065678863  0.011085152 0.005974610 0.17811300
#>  [9,] 0.017503321 0.013280601 0.0110949773  0.012359050 0.011701899 0.06668698
#> [10,] 0.031036560 0.009829463 0.0118831740  0.025204370 0.012031116 0.16081594
#> [11,] 0.004103521 0.004704373 0.0036274805  0.002476137 0.001586472 0.09900097
#> [12,] 0.006811687 0.004421187 0.0010926618  0.002526063 0.004594168 0.06619071
#> [13,] 0.003795667 0.010462458 0.0047477341  0.004456724 0.020830267 0.07039215
#> [14,] 0.021523611 0.016548959 0.0144541423  0.012825051 0.010564796 0.10138787
#> [15,] 0.017188314 0.005879904 0.0039694518  0.004529443 0.008674634 0.04966667
#> [16,] 0.011680711 0.009074380 0.0031723543  0.001143265 0.005561323 0.07173762
#> [17,] 0.005933988 0.003681387 0.0029372820  0.001664683 0.002616540 0.09461623
#> [18,] 0.008397409 0.010061908 0.0079403298  0.003940542 0.005229990 0.11148028
#> [19,] 0.012223872 0.009135698 0.0079525148  0.005787582 0.002922197 0.16931059
#> [20,] 0.011851622 0.008332394 0.0065099677  0.003428554 0.007628781 0.06549847
#> [21,] 0.004068991 0.009154063 0.0061438925  0.002919862 0.004335567 0.08083286
#> [22,] 0.007800000 0.007500000 0.0105000000  0.016000000 0.009500000 0.08160000
#> [23,] 0.009825430 0.001745934 0.0072368477 -0.002309006 0.002435790 0.11007166
#> [24,] 0.005279403 0.004913343 0.0060324736  0.003571588 0.007241577 0.05098355
#> [25,] 0.008640557 0.007058317 0.0070862684  0.006082854 0.004202256 0.06393233
#>               V60         V7         V8          V9
#>  [1,] 0.004515319 0.13697431 0.18965970  0.31759140
#>  [2,] 0.005963919 0.12134018 0.12291238  0.08627296
#>  [3,] 0.009550381 0.15372514 0.10655485  0.06574957
#>  [4,] 0.008231897 0.10836435 0.08526384  0.10825503
#>  [5,] 0.004399452 0.12357090 0.13001650  0.17130429
#>  [6,] 0.007918232 0.04041540 0.08657919  0.23328976
#>  [7,] 0.007018412 0.16035347 0.23362720  0.37325681
#>  [8,] 0.003903849 0.15596838 0.17459893  0.20843309
#>  [9,] 0.006780114 0.05617147 0.11106923  0.24535248
#> [10,] 0.004945593 0.27627495 0.31645651  0.31586380
#> [11,] 0.004442896 0.12678886 0.14914904  0.19749245
#> [12,] 0.004797165 0.05655340 0.06327169  0.10077406
#> [13,] 0.016248711 0.01166884 0.02848329  0.06402262
#> [14,] 0.010842588 0.13067156 0.11183329  0.13864291
#> [15,] 0.009735051 0.03528175 0.06801978  0.09630696
#> [16,] 0.004535896 0.03771402 0.02631021 -0.07432342
#> [17,] 0.002999699 0.10190408 0.07998767  0.15447027
#> [18,] 0.004740279 0.11919017 0.10566216  0.12139510
#> [19,] 0.014922961 0.13356613 0.14137215  0.02766207
#> [20,] 0.003591703 0.10436311 0.10571229  0.11349468
#> [21,] 0.003331873 0.12873744 0.10335004  0.11815248
#> [22,] 0.001100000 0.09930000 0.03150000  0.07360000
#> [23,] 0.001800418 0.15653088 0.13031376  0.10140175
#> [24,] 0.004510797 0.05341048 0.05010105  0.07742504
#> [25,] 0.005328822 0.08927875 0.10064257  0.13646084
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913