Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12       V13       V14
#>  [1,] 0.009510781 0.36595051 0.42070063 0.45689700 0.5349771 0.4284025
#>  [2,] 0.023413508 0.26387935 0.32405682 0.36691806 0.3225807 0.2626932
#>  [3,] 0.029622870 0.27647236 0.23845866 0.26193966 0.2997547 0.2525713
#>  [4,] 0.008439330 0.20314158 0.11378960 0.05719222 0.1992930 0.3011969
#>  [5,] 0.092287054 0.55947017 0.41921869 0.16092896 0.1505066 0.2835072
#>  [6,] 0.040931649 0.29782680 0.34265834 0.36957650 0.3058040 0.1770569
#>  [7,] 0.055135957 0.24278849 0.37419508 0.39484024 0.4284879 0.4143917
#>  [8,] 0.065400000 0.34220000 0.21280000 0.13770000 0.4032000 0.5684000
#>  [9,] 0.029560445 0.22314667 0.26436470 0.31068549 0.3392748 0.3043404
#> [10,] 0.026285279 0.45274646 0.52712061 0.52391487 0.4809860 0.3707201
#> [11,] 0.024907193 0.19824282 0.22633983 0.24859585 0.2692675 0.3551898
#> [12,] 0.019363088 0.21957013 0.31803450 0.31634837 0.2845630 0.2348539
#> [13,] 0.021456392 0.16478932 0.23902831 0.33039726 0.4356514 0.5565825
#> [14,] 0.035859488 0.06005894 0.18357603 0.27651208 0.2844252 0.2991760
#> [15,] 0.019563518 0.11164318 0.13962783 0.11983629 0.1339779 0.1091514
#> [16,] 0.006155575 0.18293641 0.12027255 0.15021128 0.3047793 0.2073293
#> [17,] 0.029835006 0.09081159 0.14898909 0.05758363 0.1003023 0.1096859
#> [18,] 0.010188468 0.16679712 0.18543212 0.13049544 0.1931089 0.0881366
#> [19,] 0.021295825 0.32482000 0.33654177 0.31703850 0.4035272 0.5932443
#> [20,] 0.021674992 0.12966983 0.12306562 0.11632544 0.2075875 0.3410741
#> [21,] 0.019106995 0.18748300 0.16862564 0.22565318 0.3252752 0.1444943
#> [22,] 0.002713372 0.02218482 0.03085014 0.07987580 0.1293906 0.1130281
#> [23,] 0.023636582 0.10311747 0.05817532 0.19423429 0.1767982 0.1093062
#> [24,] 0.065643648 0.29514989 0.27523385 0.08816997 0.1465490 0.4747375
#> [25,] 0.014950841 0.13528106 0.13258653 0.15566032 0.1391785 0.1293805
#>              V15        V16       V17       V18        V19          V2
#>  [1,] 0.33590905 0.34907172 0.3387403 0.2861134 0.40931577 0.031834995
#>  [2,] 0.23890311 0.29309438 0.3388988 0.4616238 0.62571003 0.042171382
#>  [3,] 0.25293055 0.28927030 0.2712168 0.1891314 0.27910087 0.037695641
#>  [4,] 0.42119660 0.60218754 0.6618633 0.7295942 0.82509044 0.005576037
#>  [5,] 0.44516129 0.44620530 0.2382488 0.2023353 0.37402364 0.099615079
#>  [6,] 0.14277018 0.17948147 0.1991104 0.2527162 0.26359857 0.046602860
#>  [7,] 0.57940338 0.63374712 0.7261553 0.7831392 0.81693536 0.082156843
#>  [8,] 0.23980000 0.43310000 0.5954000 0.5772000 0.81760000 0.064900000
#>  [9,] 0.35076809 0.42713229 0.4669187 0.4835324 0.49688680 0.039858105
#> [10,] 0.26764571 0.25854665 0.2787403 0.3278098 0.51642355 0.035899264
#> [11,] 0.43203564 0.52050325 0.6118062 0.6877171 0.72442811 0.012870785
#> [12,] 0.18277235 0.12981091 0.2607365 0.3226066 0.38484856 0.033123771
#> [13,] 0.63388482 0.61427481 0.6787184 0.7662866 0.85034134 0.045517558
#> [14,] 0.39821425 0.49725069 0.6591484 0.7188328 0.78041613 0.037193376
#> [15,] 0.05523161 0.07552815 0.1115450 0.1873638 0.38891333 0.012366818
#> [16,] 0.09358127 0.28166857 0.3554232 0.3003464 0.49578774 0.031633878
#> [17,] 0.14310109 0.17286863 0.1587877 0.1269432 0.06610929 0.021985455
#> [18,] 0.21252520 0.46777944 0.7166412 0.8393297 0.73277383 0.016293287
#> [19,] 0.73316379 0.82592293 0.8518860 0.7584848 0.56726653 0.030633445
#> [20,] 0.51375292 0.68118306 0.8016881 0.8285019 0.85288380 0.020094836
#> [21,] 0.22783433 0.27156329 0.3118280 0.3438800 0.29152761 0.049882897
#> [22,] 0.05754536 0.08233447 0.1115493 0.1369430 0.11649451 0.005478973
#> [23,] 0.18028231 0.21352207 0.1712454 0.1203507 0.12271347 0.028558375
#> [24,] 0.66448423 0.68831461 0.7389721 0.5984981 0.40071022 0.062624319
#> [25,] 0.09150017 0.14988043 0.2869992 0.4024333 0.41738938 0.014149534
#>             V20       V21       V22       V23       V24        V25       V26
#>  [1,] 0.6879467 0.7703583 0.6824495 0.5806715 0.6098823 0.66202744 0.8621978
#>  [2,] 0.7313557 0.8372000 0.8427736 0.8247448 0.9436947 0.96623440 0.9588304
#>  [3,] 0.3101571 0.3239194 0.3402346 0.4242781 0.4763245 0.55945463 0.6279510
#>  [4,] 0.8439495 0.9283287 0.9484576 0.8812813 0.6020409 0.37999733 0.4113401
#>  [5,] 0.8704483 0.9539593 0.4195544 0.1118556 0.1684765 0.18173452 0.3577273
#>  [6,] 0.2489721 0.3686098 0.5832435 0.7255110 0.8127542 0.87506205 0.9160591
#>  [7,] 0.7847035 0.8468827 0.8131273 0.6974162 0.6235142 0.49318152 0.3072606
#>  [8,] 0.8835000 0.5248000 0.6373000 0.8375000 0.6699000 0.77560000 0.8750000
#>  [9,] 0.6053078 0.7417723 0.8721147 0.9620450 0.9950776 0.99996435 0.9980606
#> [10,] 0.7102832 0.8078735 0.8405848 0.7754841 0.8037528 0.79792879 0.8286075
#> [11,] 0.7821189 0.8557869 0.9064153 0.9373087 0.9832610 0.98327524 0.9300541
#> [12,] 0.4077460 0.4884249 0.4843965 0.4913375 0.5604738 0.65632434 0.7645050
#> [13,] 0.9137375 0.9857230 0.9959658 0.8914115 0.8769327 0.87931331 0.8235012
#> [14,] 0.8818972 0.8811314 0.8542311 0.8677191 0.9586406 0.95910505 0.7813904
#> [15,] 0.3766258 0.2590303 0.2942381 0.4496261 0.6088227 0.60336123 0.5849839
#> [16,] 0.5508112 0.5442590 0.4500816 0.4544040 0.3831672 0.64119521 0.8779812
#> [17,] 0.1622320 0.3316762 0.5904596 0.8786328 0.9015926 0.85318139 0.8690959
#> [18,] 0.4932270 0.4947499 0.2894120 0.4001432 0.7331440 0.79540905 0.7612238
#> [19,] 0.4896795 0.5027675 0.6093942 0.6571132 0.6729311 0.59375559 0.5806721
#> [20,] 0.8308726 0.8641274 0.8895031 0.8657721 0.8449900 0.84042254 0.8744507
#> [21,] 0.1340740 0.1735135 0.1314905 0.1735464 0.1767193 0.24773036 0.2597051
#> [22,] 0.1079443 0.1949076 0.3837781 0.4366009 0.5667204 0.66067793 0.7436563
#> [23,] 0.1915420 0.2577914 0.2010231 0.2587809 0.2149690 0.07647532 0.2461227
#> [24,] 0.2954747 0.5193961 0.7560887 0.8545307 0.9548214 0.90774270 0.6524994
#> [25,] 0.5086196 0.6341064 0.7167413 0.7732466 0.7843387 0.78463709 0.7814997
#>             V27       V28       V29         V3       V30       V31       V32
#>  [1,] 0.9838514 0.9440483 0.8473597 0.05544171 0.6117987 0.5453995 0.6838035
#>  [2,] 0.9689910 0.9731207 0.8540710 0.03514542 0.6172542 0.4201764 0.3064985
#>  [3,] 0.6609932 0.7941948 0.8404924 0.04648261 0.8726850 0.8737235 0.7302210
#>  [4,] 0.6289208 0.8338012 0.8186858 0.01430859 0.6275039 0.3336408 0.3342121
#>  [5,] 0.2380410 0.4446666 0.5080553 0.07424108 0.7060116 0.3392951 0.3721637
#>  [6,] 0.9385562 0.9793584 0.9204499 0.04089258 0.7181212 0.7162956 0.5639769
#>  [7,] 0.1914141 0.1878734 0.2103011 0.09682141 0.3298704 0.3068111 0.2786080
#>  [8,] 0.8300000 0.6896000 0.3372000 0.07370000 0.6405000 0.7138000 0.8202000
#>  [9,] 0.9513684 0.9207488 0.7427371 0.01566408 0.5345525 0.4807230 0.4809548
#> [10,] 0.8998450 0.8709435 0.7029136 0.05007027 0.4957638 0.4304512 0.4447309
#> [11,] 0.8799087 0.8105038 0.6895449 0.02869856 0.5550941 0.4095118 0.3066533
#> [12,] 0.9499210 0.9938332 0.8682703 0.03788033 0.7030361 0.6063385 0.5756698
#> [13,] 0.7547971 0.6557233 0.6103088 0.03137461 0.4367736 0.2905875 0.1683196
#> [14,] 0.5424730 0.3328458 0.2675787 0.06314226 0.2853652 0.3162456 0.3807328
#> [15,] 0.5067973 0.7023572 0.9823341 0.01710332 0.9980847 0.9496421 0.8826154
#> [16,] 0.8829407 0.5402143 0.3236408 0.05204706 0.2498915 0.2894103 0.4474665
#> [17,] 0.8337655 0.8650730 0.9637178 0.02608364 0.8022666 0.6349716 0.5188153
#> [18,] 0.7717327 0.8311986 0.5970689 0.03064926 0.3895838 0.4108328 0.4626187
#> [19,] 0.6422834 0.5924046 0.4781846 0.04715373 0.5183549 0.5245669 0.3674064
#> [20,] 0.8239019 0.7047766 0.5776456 0.03351301 0.4896297 0.3905862 0.3125509
#> [21,] 0.1353438 0.5241768 0.7102164 0.03874849 0.6483777 0.5277458 0.6335371
#> [22,] 0.7533872 0.8497886 0.9821002 0.01447833 0.9232212 0.7317271 0.7038352
#> [23,] 0.5527646 0.7169546 0.7194827 0.01992733 0.7083637 0.6980795 0.5413245
#> [24,] 0.6447572 0.6363966 0.4830446 0.06733828 0.5499558 0.4620613 0.2005462
#> [25,] 0.7468034 0.7307025 0.7638982 0.02405992 0.6960071 0.6516478 0.6156555
#>              V33       V34        V35        V36        V37       V38
#>  [1,] 0.58686489 0.3191280 0.04927126 0.31784546 0.35443955 0.3908169
#>  [2,] 0.27662492 0.2021047 0.13505035 0.01899392 0.06289585 0.1230194
#>  [3,] 0.57822482 0.4746120 0.40298427 0.36292060 0.34745884 0.3786099
#>  [4,] 0.32313386 0.2102069 0.18115034 0.14685382 0.12221640 0.1722474
#>  [5,] 0.57275930 0.7241442 0.84832480 0.84106563 0.80217160 0.8884813
#>  [6,] 0.41682334 0.2253598 0.11484699 0.06414528 0.18346600 0.1645878
#>  [7,] 0.36943407 0.5072793 0.55932974 0.38424363 0.30548083 0.3580174
#>  [8,] 0.66570000 0.5254000 0.29600000 0.07040000 0.09700000 0.3941000
#>  [9,] 0.34191490 0.2237206 0.15307333 0.18692330 0.31759544 0.3981186
#> [10,] 0.44210498 0.3763423 0.19688212 0.33190475 0.33453482 0.3487745
#> [11,] 0.20998442 0.1117230 0.14106020 0.17652716 0.24743344 0.2633265
#> [12,] 0.51786664 0.4519682 0.27042405 0.13533302 0.24402716 0.2187410
#> [13,] 0.09364884 0.0462125 0.09146517 0.21039447 0.23507344 0.2807694
#> [14,] 0.49846787 0.5825716 0.51674922 0.35632953 0.30974421 0.2582381
#> [15,] 0.79555961 0.6994777 0.57571164 0.46058011 0.43033890 0.1824983
#> [16,] 0.45059453 0.3665011 0.43595503 0.54533110 0.59554466 0.3902566
#> [17,] 0.56644546 0.7128468 0.70752814 0.53676542 0.28496544 0.1614581
#> [18,] 0.34600929 0.1297517 0.26476543 0.37420037 0.24224184 0.1395982
#> [19,] 0.23301628 0.1698763 0.36069093 0.52547876 0.52796260 0.3490670
#> [20,] 0.27516552 0.3033817 0.24545785 0.14197007 0.18192127 0.1543128
#> [21,] 0.71276629 0.6349373 0.60044690 0.76282694 0.87445911 0.9373891
#> [22,] 0.77684899 0.5997914 0.48753424 0.42744444 0.36806107 0.4147428
#> [23,] 0.65488662 0.8919203 0.96357058 0.91671553 0.70290170 0.5729999
#> [24,] 0.23643720 0.1529792 0.48453283 0.72110870 0.74736370 0.3966528
#> [25,] 0.63615546 0.6259693 0.63111622 0.62143823 0.53845994 0.4457748
#>              V39         V4       V40       V41        V42        V43
#>  [1,] 0.57769138 0.04408639 0.4385479 0.3003778 0.44257749 0.43563373
#>  [2,] 0.11543769 0.04482168 0.1234391 0.0878605 0.12247016 0.20662012
#>  [3,] 0.40871745 0.04897553 0.4437761 0.5438199 0.50546965 0.47645006
#>  [4,] 0.17656857 0.05366572 0.1819305 0.1995042 0.20087320 0.16222556
#>  [5,] 0.78214988 0.08584307 0.5441496 0.3402292 0.25280129 0.23124239
#>  [6,] 0.19067768 0.06898282 0.1499080 0.1914700 0.21465173 0.17482484
#>  [7,] 0.35336523 0.10698829 0.1804542 0.1298867 0.25191129 0.26814648
#>  [8,] 0.60280000 0.11320000 0.3521000 0.3924000 0.48080000 0.46020000
#>  [9,] 0.38936480 0.02361238 0.3737244 0.2725897 0.17534767 0.21650120
#> [10,] 0.48007362 0.05120024 0.4069954 0.1877947 0.31666289 0.38869140
#> [11,] 0.24091322 0.05723147 0.2685296 0.2820124 0.27929381 0.24766183
#> [12,] 0.21379057 0.06760645 0.3412840 0.3742236 0.35750081 0.26215124
#> [13,] 0.36256059 0.01306426 0.2840555 0.1721267 0.18160253 0.13069861
#> [14,] 0.36286033 0.03318919 0.4420250 0.3543734 0.32409289 0.23461960
#> [15,] 0.05952794 0.03628146 0.1539514 0.1851222 0.18966708 0.26179776
#> [16,] 0.54434754 0.03709039 0.5272670 0.3200076 0.26394750 0.24572995
#> [17,] 0.17158020 0.04844364 0.2639127 0.2101100 0.21279837 0.29802451
#> [18,] 0.15420016 0.04140978 0.1601230 0.1014163 0.09802034 0.06612329
#> [19,] 0.23496113 0.02140409 0.3096301 0.3201968 0.26542974 0.18976971
#> [20,] 0.10578075 0.04169752 0.1237433 0.1253949 0.10058337 0.09354734
#> [21,] 0.95532144 0.03092217 0.8932631 0.8637801 0.75829702 0.54399775
#> [22,] 0.43360372 0.03387001 0.3559654 0.1636083 0.01356702 0.04180456
#> [23,] 0.50045034 0.02397005 0.6472306 0.5959651 0.52714775 0.34204205
#> [24,] 0.12828957 0.04862117 0.3531150 0.4425570 0.41085929 0.28242697
#> [25,] 0.31607715 0.03215749 0.3086207 0.2450208 0.19544112 0.19217040
#>              V44        V45        V46        V47        V48         V49
#>  [1,] 0.48407031 0.51066405 0.19804624 0.15241219 0.20209360 0.104246411
#>  [2,] 0.21909545 0.20265425 0.17712291 0.14138315 0.09843250 0.049799334
#>  [3,] 0.47437265 0.55632599 0.48727795 0.34172676 0.22771840 0.126314339
#>  [4,] 0.22269529 0.17838542 0.14572817 0.09710769 0.04223975 0.035686268
#>  [5,] 0.28967042 0.46517196 0.36033937 0.16425212 0.15610158 0.084156092
#>  [6,] 0.11591676 0.10579021 0.14534958 0.18432364 0.15035213 0.075382911
#>  [7,] 0.16317906 0.08423344 0.03464683 0.07535651 0.05701477 0.040312560
#>  [8,] 0.41640000 0.54380000 0.56490000 0.31950000 0.24840000 0.129900000
#>  [9,] 0.23154194 0.18762633 0.16414349 0.11940413 0.06032501 0.041878837
#> [10,] 0.41777186 0.38218927 0.16611045 0.10738987 0.13370474 0.071523571
#> [11,] 0.18264135 0.11361252 0.09743574 0.13105109 0.09855603 0.059181094
#> [12,] 0.18897244 0.19348624 0.17427836 0.17900628 0.16243511 0.077257356
#> [13,] 0.17260453 0.14070947 0.11892239 0.12985090 0.10739966 0.074417396
#> [14,] 0.22118268 0.19055386 0.14631137 0.08913306 0.05466684 0.037012564
#> [15,] 0.36461836 0.35573239 0.22923699 0.09905889 0.01774149 0.014448684
#> [16,] 0.24856255 0.09889194 0.06971053 0.05083359 0.03203738 0.037876178
#> [17,] 0.17668339 0.06012022 0.06222638 0.06264931 0.04978090 0.007054083
#> [18,] 0.07454378 0.07152990 0.07846998 0.04291886 0.04214266 0.015298351
#> [19,] 0.16524429 0.14405310 0.08874119 0.10020030 0.08588865 0.059291521
#> [20,] 0.09310271 0.10654074 0.07241656 0.07801117 0.05297181 0.028413386
#> [21,] 0.29685275 0.39319842 0.48077572 0.38380591 0.26268347 0.175237319
#> [22,] 0.08642967 0.12118637 0.03928885 0.01651218 0.03834380 0.014936477
#> [23,] 0.16765699 0.14154553 0.10912096 0.11933866 0.08492611 0.040381931
#> [24,] 0.20939831 0.24417561 0.24543730 0.21515069 0.19037647 0.123346131
#> [25,] 0.18767901 0.10076412 0.07040419 0.04987792 0.04451969 0.016630496
#>               V5         V50          V51         V52         V53         V54
#>  [1,] 0.06425405 0.023838599  0.035193283 0.018703747 0.002536091 0.010652811
#>  [2,] 0.07392986 0.015290726  0.017373184 0.014412300 0.010054875 0.008929676
#>  [3,] 0.07040260 0.041824691  0.027069827 0.023188692 0.016589697 0.012326014
#>  [4,] 0.07265444 0.008875298  0.006417381 0.009371803 0.004059919 0.001701004
#>  [5,] 0.06564705 0.037318470  0.023549708 0.020229254 0.015235115 0.018974065
#>  [6,] 0.08488535 0.017287475  0.018615078 0.014038946 0.007433956 0.010141189
#>  [7,] 0.11011162 0.010764660  0.018144417 0.013665091 0.009780828 0.023284163
#>  [8,] 0.24820000 0.082500000  0.024300000 0.021000000 0.036100000 0.023900000
#>  [9,] 0.04396624 0.009509101  0.014860438 0.002840369 0.010578044 0.012237602
#> [10,] 0.07717167 0.015637868  0.022262343 0.016374696 0.009644749 0.012709881
#> [11,] 0.09981957 0.028173837  0.018284937 0.017014746 0.018510523 0.011881049
#> [12,] 0.06709234 0.011861192  0.021166433 0.014514708 0.009180879 0.010239629
#> [13,] 0.09263367 0.026626411  0.025023576 0.017748708 0.013003899 0.025211886
#> [14,] 0.09853698 0.002692970 -0.002785906 0.005618417 0.010882727 0.009119744
#> [15,] 0.03401575 0.016849337  0.011218301 0.003555339 0.006815367 0.011233342
#> [16,] 0.04951140 0.013062871  0.009363529 0.006412609 0.003468645 0.017868316
#> [17,] 0.06758500 0.008045680  0.007379994 0.003488634 0.004345453 0.006601363
#> [18,] 0.07053080 0.015369665  0.005902347 0.020176848 0.008189328 0.013166746
#> [19,] 0.04006924 0.016629127  0.013674095 0.007992218 0.004862941 0.008852235
#> [20,] 0.07767028 0.013478074  0.012259893 0.007831962 0.005875789 0.009120549
#> [21,] 0.06203866 0.069035212  0.035346857 0.016778542 0.015568118 0.010326719
#> [22,] 0.02674972 0.004447037  0.004674076 0.005629365 0.012499114 0.004787300
#> [23,] 0.03482255 0.025278503  0.018699237 0.009122408 0.015580421 0.014817659
#> [24,] 0.04818281 0.030617588  0.015775223 0.014599218 0.010794889 0.010425803
#> [25,] 0.06763844 0.013789110  0.005642452 0.004531995 0.006773016 0.007591076
#>               V55         V56         V57          V58         V59         V6
#>  [1,] 0.013698904 0.003698126 0.004922341 3.064687e-03 0.003229374 0.10274641
#>  [2,] 0.002254104 0.005656633 0.004556466 1.472647e-03 0.002226035 0.10501796
#>  [3,] 0.007666372 0.006906896 0.009835530 9.310099e-03 0.007627592 0.10934195
#>  [4,] 0.001272430 0.002967919 0.001583428 5.991085e-03 0.002856050 0.11099911
#>  [5,] 0.017598108 0.016731793 0.015285211 1.484676e-02 0.011818165 0.11627321
#>  [6,] 0.005920849 0.007186281 0.007301688 7.115474e-03 0.008590660 0.14569972
#>  [7,] 0.014891240 0.010907302 0.016442166 2.528502e-02 0.013928922 0.08520470
#>  [8,] 0.044700000 0.039400000 0.035500000 4.400000e-02 0.024300000 0.12570000
#>  [9,] 0.012238884 0.009185154 0.006383915 6.336375e-03 0.008813633 0.06792361
#> [10,] 0.010463698 0.005261714 0.006089386 6.535558e-03 0.007127647 0.12424495
#> [11,] 0.003954564 0.010901288 0.010056350 8.906828e-03 0.007640698 0.12348219
#> [12,] 0.007129279 0.006775962 0.003213981 7.113333e-03 0.007000173 0.05914259
#> [13,] 0.015406492 0.006228580 0.006856158 7.594397e-03 0.009331601 0.16778694
#> [14,] 0.007275245 0.011930209 0.007455001 7.160309e-03 0.008781353 0.14142261
#> [15,] 0.001103899 0.001761062 0.004957514 4.025371e-03 0.006629968 0.02114727
#> [16,] 0.017794818 0.006617579 0.012737828 1.687358e-02 0.015444056 0.03728229
#> [17,] 0.004999092 0.006786817 0.013655683 1.192568e-02 0.005386363 0.09543682
#> [18,] 0.011299732 0.006779417 0.007833544 4.362766e-03 0.004831658 0.05673412
#> [19,] 0.006270808 0.009542825 0.010513668 6.658843e-03 0.008915479 0.13578273
#> [20,] 0.006711709 0.008081572 0.006069147 4.409017e-03 0.004750433 0.10877948
#> [21,] 0.017494936 0.012058601 0.002818522 1.603259e-02 0.010648887 0.10271356
#> [22,] 0.005305530 0.009362314 0.003864472 8.476284e-05 0.005448480 0.02224770
#> [23,] 0.008939434 0.008660565 0.008098480 7.445428e-03 0.011563538 0.06017941
#> [24,] 0.006534424 0.005106885 0.020177450 9.362253e-03 0.005938679 0.06932973
#> [25,] 0.005408885 0.006411838 0.006627066 4.162944e-03 0.003481848 0.10222035
#>               V60         V7         V8         V9
#>  [1,] 0.003138282 0.10094235 0.14744278 0.25870406
#>  [2,] 0.003863051 0.12311807 0.14934197 0.20878452
#>  [3,] 0.004167329 0.11381481 0.14093821 0.24004197
#>  [4,] 0.004437167 0.14644073 0.15978951 0.20653053
#>  [5,] 0.011386672 0.20927064 0.35452276 0.54706967
#>  [6,] 0.008184004 0.14812666 0.17901482 0.20236982
#>  [7,] 0.006577592 0.11158738 0.06109464 0.11828234
#>  [8,] 0.009800000 0.17970000 0.09890000 0.24600000
#>  [9,] 0.004018845 0.08093963 0.11251816 0.16713156
#> [10,] 0.004601351 0.13698651 0.21867930 0.33655315
#> [11,] 0.006313444 0.14022913 0.13702206 0.16445262
#> [12,] 0.003368273 0.08503276 0.07552305 0.17322054
#> [13,] 0.007869135 0.17496158 0.14553833 0.17439141
#> [14,] 0.004393958 0.12525163 0.06072090 0.10839625
#> [15,] 0.001570615 0.08273308 0.08489091 0.07416306
#> [16,] 0.005698055 0.03908202 0.01363734 0.06679369
#> [17,] 0.005059317 0.08611841 0.11290455 0.07567091
#> [18,] 0.003281003 0.04192472 0.01367432 0.09329806
#> [19,] 0.005760065 0.17357777 0.26420961 0.29114430
#> [20,] 0.004742925 0.11366263 0.08478602 0.12575063
#> [21,] 0.012503744 0.21035746 0.21804906 0.04462130
#> [22,] 0.003615427 0.03075079 0.01736984 0.01385737
#> [23,] 0.008379557 0.06276827 0.08289561 0.05927989
#> [24,] 0.004720695 0.19195388 0.27634680 0.27983371
#> [25,] 0.004995338 0.11962725 0.10863844 0.13904666
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841