Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1        V10        V11        V12        V13        V14
#>  [1,]  0.038658791 0.46217748 0.40674455 0.29508616 0.26395813 0.17993636
#>  [2,]  0.018359922 0.21387826 0.30332972 0.30394807 0.27992134 0.24536095
#>  [3,]  0.031353101 0.31849875 0.25072406 0.29554523 0.28226143 0.23962689
#>  [4,]  0.065387213 0.03158326 0.11931165 0.19837561 0.24561387 0.37523252
#>  [5,]  0.022287219 0.29308315 0.44610699 0.49340038 0.42123974 0.33710551
#>  [6,]  0.020171055 0.17002609 0.10303855 0.03913140 0.12399232 0.24033893
#>  [7,]  0.008821636 0.19249622 0.29411227 0.32759857 0.30073516 0.21658973
#>  [8,]  0.062900000 0.35970000 0.54660000 0.52050000 0.51270000 0.53950000
#>  [9,]  0.046102415 0.23817831 0.33935384 0.46271280 0.53346037 0.53676851
#> [10,]  0.030471270 0.30556740 0.34068829 0.34655271 0.27067096 0.14625465
#> [11,]  0.022306900 0.23457872 0.29604600 0.31308934 0.41931254 0.57548416
#> [12,]  0.026699464 0.21292621 0.26651514 0.32001471 0.33244309 0.36295113
#> [13,]  0.034036485 0.40358186 0.48267869 0.49003427 0.46718687 0.36394539
#> [14,]  0.016548551 0.17733640 0.13595592 0.07397234 0.07659462 0.12266138
#> [15,]  0.017693970 0.02944700 0.05206419 0.20855085 0.15635205 0.09704061
#> [16,]  0.019012127 0.28497065 0.34552118 0.39151593 0.42641594 0.54265277
#> [17,]  0.024981868 0.13774847 0.17189471 0.21468000 0.19688532 0.18331972
#> [18,]  0.010627411 0.07809725 0.07016458 0.05705886 0.08798212 0.08800998
#> [19,]  0.023586583 0.07959011 0.06767588 0.08416109 0.18645565 0.35017138
#> [20,]  0.014300146 0.10126206 0.12585710 0.15643693 0.17087162 0.18750905
#> [21,]  0.010583320 0.03745932 0.10908981 0.14037924 0.22229019 0.29847329
#> [22,]  0.021544352 0.27942476 0.25885554 0.10384419 0.21886023 0.49640541
#> [23,]  0.037331710 0.19631162 0.30393849 0.30786496 0.36249126 0.38037132
#> [24,]  0.007911681 0.14162574 0.12528576 0.11977351 0.15405177 0.13372524
#> [25,] -0.003288395 0.17045421 0.10620581 0.16362591 0.35254431 0.20927048
#>               V15        V16       V17       V18       V19           V2
#>  [1,] 0.225285970 0.42163423 0.3882197 0.4102824 0.3424003  0.025218610
#>  [2,] 0.197074901 0.18285317 0.2375239 0.2591520 0.3861723  0.040048149
#>  [3,] 0.114811589 0.12839503 0.1331844 0.1783509 0.1550379  0.040650648
#>  [4,] 0.488839261 0.56614296 0.6855489 0.7403652 0.8057802  0.078971550
#>  [5,] 0.259637593 0.37691060 0.4054244 0.3959530 0.3741327  0.028290791
#>  [6,] 0.415673779 0.59934959 0.7083327 0.7538859 0.8462736  0.018402658
#>  [7,] 0.069619912 0.24143682 0.4239060 0.6020417 0.7517152 -0.004329085
#>  [8,] 0.655800000 0.87050000 0.9786000 0.9335000 0.7917000  0.106500000
#>  [9,] 0.619393572 0.74659493 0.9003300 0.9220428 0.9142502  0.085435021
#> [10,] 0.148726133 0.19831766 0.2074456 0.2627683 0.3105481  0.039058931
#> [11,] 0.648171469 0.70881667 0.7393682 0.8174380 0.8950150  0.033613674
#> [12,] 0.435292284 0.52580904 0.5866357 0.6617290 0.7192740  0.031586644
#> [13,] 0.250329426 0.23882868 0.2583860 0.3071468 0.4597155  0.045290221
#> [14,] 0.144858347 0.17004107 0.2290965 0.2414758 0.3354398  0.001939324
#> [15,] 0.083459760 0.11664446 0.1351009 0.1401190 0.1309264  0.030561550
#> [16,] 0.711269218 0.83290385 0.9664480 0.9364944 0.8635804  0.029979288
#> [17,] 0.233994969 0.47800574 0.6596550 0.5630309 0.3800264  0.026006931
#> [18,] 0.003458583 0.09224146 0.1509074 0.3730527 0.3291696  0.005124663
#> [19,] 0.515730127 0.70189968 0.7787099 0.7875130 0.8149520  0.026206337
#> [20,] 0.159074967 0.16866599 0.1988020 0.2664104 0.2699955  0.017635382
#> [21,] 0.413232759 0.47707216 0.5471922 0.6366832 0.7144067  0.007617280
#> [22,] 0.658148168 0.77417221 0.8426777 0.7310157 0.5098652  0.053287193
#> [23,] 0.566666787 0.75601166 0.9435380 0.9853991 0.9788553  0.033009167
#> [24,] 0.181654682 0.22663619 0.2839205 0.2793029 0.2273563 -0.001057317
#> [25,] 0.002500146 0.18215178 0.3270514 0.3662195 0.6789255  0.035824500
#>              V20        V21        V22       V23       V24        V25       V26
#>  [1,] 0.38464614 0.28551094 0.10235245 0.1920010 0.3488304 0.50369212 0.6656755
#>  [2,] 0.40118833 0.44391651 0.43189777 0.4085599 0.5092836 0.60546659 0.7590039
#>  [3,] 0.15602692 0.05979147 0.21946326 0.4783312 0.6132617 0.68181624 0.7696038
#>  [4,] 0.84989679 0.83990463 0.85493028 0.8631466 0.8340107 0.73305607 0.5788242
#>  [5,] 0.48103217 0.74404091 0.99566583 0.9981863 0.9237582 0.95590423 0.8702771
#>  [6,] 0.93677991 0.95450653 1.01843029 0.8789654 0.5957624 0.43362722 0.3560087
#>  [7,] 0.73965075 0.78122901 0.84883359 0.9335438 0.9630092 1.05158405 0.9074699
#>  [8,] 0.73830000 0.69080000 0.38500000 0.0671000 0.0502000 0.27170000 0.2839000
#>  [9,] 0.83984938 0.76658299 0.60869110 0.4544071 0.3900233 0.26185568 0.1503945
#> [10,] 0.35553427 0.51110608 0.70353967 0.7928731 0.8558000 0.90767101 0.9750852
#> [11,] 0.92776538 0.93969872 0.94169079 0.8993626 0.8079252 0.70499190 0.6942883
#> [12,] 0.78207500 0.81658459 0.88144707 0.9227224 0.9660676 0.94936780 0.9305663
#> [13,] 0.68055672 0.78957322 0.77868906 0.7135311 0.7125099 0.67459124 0.7793854
#> [14,] 0.52985017 0.64577669 0.82958892 0.9752548 0.9051103 0.86755375 0.9384558
#> [15,] 0.07181114 0.16192021 0.02958336 0.1725944 0.1763385 0.02852985 0.1958683
#> [16,] 0.78881247 0.74657981 0.80851727 0.6729236 0.5635032 0.66473251 0.7736334
#> [17,] 0.37845000 0.41906561 0.46055093 0.5246196 0.6788783 0.80581526 0.8107655
#> [18,] 0.26629340 0.16411570 0.19912106 0.2094718 0.2196207 0.05401203 0.1841063
#> [19,] 0.85772171 0.97435472 0.92754623 0.8928402 0.9188978 0.90500689 0.9465662
#> [20,] 0.35906162 0.42968766 0.51263691 0.6494418 0.7091976 0.76899494 0.7978691
#> [21,] 0.79740741 0.95221971 1.00000000 0.9665661 0.9487658 0.87096311 0.7942210
#> [22,] 0.35217398 0.50325753 0.70876606 0.8226958 0.9576452 0.82764576 0.7017542
#> [23,] 0.86163604 0.83891117 0.86343666 0.5728250 0.4173487 0.57525332 0.6561581
#> [24,] 0.28545960 0.37195548 0.43627305 0.5505769 0.6739652 0.86742017 0.9089202
#> [25,] 0.68056113 0.63273606 0.52593651 0.4792514 0.3674195 0.64418825 0.9167589
#>             V27       V28       V29           V3       V30       V31       V32
#>  [1,] 0.7835478 0.9041905 0.9331459  0.014052270 1.0298522 0.8948412 0.7222383
#>  [2,] 0.9250231 0.9838109 0.9033922  0.033345432 0.7459424 0.5968514 0.5340289
#>  [3,] 0.9122456 0.9914965 0.9178655  0.031791981 0.9097130 0.9343593 0.9032437
#>  [4,] 0.5259986 0.4314458 0.4366358  0.127129478 0.3341230 0.1707894 0.1690772
#>  [5,] 0.7913273 0.6957956 0.4653797  0.024680563 0.3293591 0.2525973 0.4063798
#>  [6,] 0.2424012 0.2976159 0.5241648  0.006120974 0.5695009 0.3866749 0.3403661
#>  [7,] 0.5107007 0.3825093 0.3449418  0.002146453 0.5371282 0.6927709 0.8147465
#>  [8,] 0.2234000 0.1911000 0.0408000  0.152600000 0.2531000 0.1979000 0.1891000
#>  [9,] 0.1552536 0.1497513 0.1657024  0.088689501 0.1775255 0.2497986 0.2986447
#> [10,] 0.9963301 0.9744536 0.8370096  0.038314635 0.6419017 0.5837485 0.4926152
#> [11,] 0.6523922 0.5780669 0.5185199  0.033221376 0.4208773 0.3211257 0.1982279
#> [12,] 0.9048685 0.8537487 0.7426333  0.030089152 0.5849329 0.4328076 0.3275068
#> [13,] 0.9065420 0.9040238 0.7366372  0.059491388 0.5456301 0.4519741 0.4751872
#> [14,] 0.9463432 0.8927162 0.8107371  0.003799013 0.7137695 0.6636426 0.5585078
#> [15,] 0.4104805 0.6396396 0.6765730  0.027102853 0.6451422 0.7060493 0.5939737
#> [16,] 0.5979187 0.4727329 0.6073526  0.032280585 0.7106454 0.6005526 0.4128784
#> [17,] 0.7868364 0.8225248 0.7827395  0.031936927 0.7338520 0.7741398 0.7190491
#> [18,] 0.4154291 0.6568615 0.8285253 -0.008559054 0.7726637 0.5750654 0.2143976
#> [19,] 0.9030875 0.7039333 0.3991392  0.023795519 0.3460417 0.3723202 0.3024364
#> [20,] 0.7494021 0.7653379 0.8043434  0.021058161 0.7705171 0.6871952 0.6198166
#> [21,] 0.6518378 0.4459520 0.4748586  0.005195923 0.6242627 0.5014993 0.3215749
#> [22,] 0.7099176 0.5502279 0.4130417  0.089544114 0.4719292 0.4329505 0.2201616
#> [23,] 0.6372854 0.5214967 0.5627695  0.044180150 0.4842742 0.4805953 0.3551586
#> [24,] 0.8375272 0.8073403 0.6498875 -0.007551815 0.4390873 0.3684858 0.2818883
#> [25,] 0.9608243 0.5087827 0.3013544  0.058792759 0.1295319 0.1579136 0.4323358
#>             V33       V34       V35        V36        V37        V38        V39
#>  [1,] 0.6409859 0.6368579 0.7019539 0.75103983 0.81507010 0.77090681 0.63873046
#>  [2,] 0.4505111 0.3734528 0.2546429 0.17513489 0.23954213 0.19606786 0.15775535
#>  [3,] 0.7592672 0.6489931 0.6318338 0.62818366 0.63290413 0.51603493 0.42800024
#>  [4,] 0.2580419 0.3538994 0.5036320 0.47533911 0.38923264 0.40536712 0.41946368
#>  [5,] 0.4127914 0.3262176 0.1392346 0.23230903 0.32918475 0.30662057 0.44778206
#>  [6,] 0.4299109 0.3587534 0.2256414 0.16506298 0.12071049 0.15971316 0.14617252
#>  [7,] 0.9455933 0.8969819 0.6622921 0.38808043 0.04103697 0.08157908 0.33351299
#>  [8,] 0.2433000 0.1956000 0.2667000 0.13400000 0.10730000 0.20230000 0.17940000
#>  [9,] 0.3215067 0.4022994 0.4360946 0.27240343 0.24047198 0.31216655 0.28384485
#> [10,] 0.3951788 0.2107606 0.1215371 0.01596556 0.07317934 0.11702061 0.12326457
#> [11,] 0.1285959 0.1472891 0.1973974 0.25372520 0.23375692 0.23659270 0.28932082
#> [12,] 0.2620122 0.1505218 0.1425825 0.12236915 0.19614844 0.26944765 0.24058130
#> [13,] 0.4305613 0.3601775 0.2191900 0.30359229 0.30503014 0.35428153 0.52356587
#> [14,] 0.5001963 0.5304636 0.5243534 0.45013028 0.43706559 0.32784467 0.16868452
#> [15,] 0.6773229 0.8566832 0.9297711 0.95484514 0.81043243 0.71110235 0.69173834
#> [16,] 0.3764663 0.3743142 0.1966895 0.18759515 0.26401923 0.20819135 0.17798381
#> [17,] 0.6791477 0.7394839 0.7524341 0.72777233 0.69461575 0.58670119 0.40738755
#> [18,] 0.4761150 0.4458965 0.2022507 0.29576406 0.43296552 0.55062787 0.65723933
#> [19,] 0.2575654 0.2275103 0.1692596 0.07723152 0.16439984 0.14852315 0.09425939
#> [20,] 0.6799645 0.6651094 0.6251773 0.51942158 0.36082602 0.27301691 0.26118534
#> [21,] 0.3766233 0.3321971 0.1002426 0.06479094 0.03560188 0.04236391 0.16880249
#> [22,] 0.1868771 0.1145605 0.4569366 0.68205821 0.64183330 0.33582204 0.14191289
#> [23,] 0.3979834 0.3709899 0.2326092 0.12904144 0.15964688 0.10237778 0.25389456
#> [24,] 0.2051064 0.1932644 0.2522279 0.25688119 0.17993577 0.15596235 0.16076574
#> [25,] 0.5004581 0.3440313 0.3038303 0.39569032 0.54008440 0.35067297 0.48739266
#>                 V4       V40       V41        V42        V43          V44
#>  [1,]  0.066662791 0.5727440 0.4046130 0.28914445 0.31306977  0.455479694
#>  [2,]  0.047492706 0.2652847 0.3841675 0.43910816 0.37715452  0.183691315
#>  [3,]  0.034138162 0.5091979 0.5407677 0.49832389 0.46523733  0.520695169
#>  [4,]  0.173554109 0.4369769 0.4245307 0.43606774 0.41740209  0.324343235
#>  [5,]  0.040702401 0.4513203 0.3276611 0.08816202 0.30751440  0.341329077
#>  [6,]  0.042966635 0.1085653 0.1704705 0.16057247 0.10162219  0.156074606
#>  [7,]  0.020305429 0.5595110 0.5586593 0.42786668 0.22302478  0.072761845
#>  [8,]  0.122900000 0.0227000 0.1313000 0.17750000 0.15490000  0.162600000
#>  [9,]  0.111914434 0.2007994 0.1900210 0.16323740 0.13283793  0.072770513
#> [10,]  0.045137393 0.0904187 0.1021398 0.07975686 0.16689989  0.180577014
#> [11,]  0.055067331 0.2755494 0.2265986 0.20800665 0.21831198  0.193213120
#> [12,]  0.061321366 0.2303849 0.2819146 0.26548397 0.18876512  0.164508375
#> [13,]  0.049532343 0.4390064 0.2685263 0.38756287 0.40583284  0.395795325
#> [14,]  0.030126193 0.1111146 0.1111277 0.02204049 0.04562247 -0.009184494
#> [15,]  0.019597434 0.7029027 0.6233581 0.54054178 0.39084199  0.215102778
#> [16,]  0.039766374 0.2869771 0.2202142 0.21554142 0.12127564  0.127853868
#> [17,]  0.041960962 0.2830269 0.2315736 0.21353703 0.17487138  0.178826797
#> [18,]  0.001006708 0.5389022 0.2433057 0.22994426 0.17777899  0.079311746
#> [19,]  0.039521908 0.1232041 0.0950800 0.04567870 0.06902573  0.123351672
#> [20,]  0.028746003 0.3274498 0.2951160 0.27489627 0.24752260  0.184681869
#> [21,]  0.045216451 0.2169635 0.1633292 0.06138180 0.01454520  0.077478263
#> [22,]  0.072497110 0.3273273 0.3824593 0.34308451 0.23407909  0.195371971
#> [23,]  0.038361905 0.3085971 0.2581203 0.26870209 0.16642139  0.059301633
#> [24,] -0.011223683 0.1451036 0.1754426 0.19377137 0.16256196  0.121442188
#> [25,]  0.031288551 0.4526321 0.2675609 0.28138575 0.25318084  0.241804117
#>               V45         V46        V47        V48        V49         V5
#>  [1,]  0.45273353 0.238492648 0.09623565 0.15645954 0.09212215 0.07731287
#>  [2,]  0.21194485 0.196906991 0.19026840 0.16440344 0.08760311 0.06412175
#>  [3,]  0.55658774 0.464952741 0.34955906 0.24511351 0.12230243 0.05302613
#>  [4,]  0.23500404 0.108079720 0.11484591 0.07938110 0.03881705 0.19288260
#>  [5,]  0.39234773 0.339568747 0.20218228 0.09178474 0.06582510 0.03682946
#>  [6,]  0.16949965 0.073012794 0.07402312 0.06681021 0.03770335 0.07717537
#>  [7,]  0.08825615 0.119379042 0.14052595 0.13575528 0.09144604 0.03357731
#>  [8,]  0.07080000 0.012900000 0.07950000 0.07620000 0.01170000 0.14370000
#>  [9,]  0.06763767 0.066161269 0.08073758 0.07870710 0.05622976 0.13116680
#> [10,]  0.13836676 0.130209161 0.13534866 0.12055506 0.07134386 0.06194686
#> [11,]  0.12889058 0.094990262 0.09840465 0.09878140 0.06498770 0.11963592
#> [12,]  0.14660385 0.094409759 0.07805188 0.06620720 0.04215707 0.10017616
#> [13,]  0.40510104 0.212463265 0.13319805 0.14414964 0.09052302 0.08880650
#> [14,] -0.03200460 0.007319573 0.02310064 0.02425143 0.01282485 0.05246384
#> [15,]  0.17274875 0.127426703 0.12355159 0.10567633 0.07258398 0.04800008
#> [16,]  0.18626007 0.193946254 0.17424573 0.12822590 0.06192125 0.07674864
#> [17,]  0.15095419 0.147027115 0.09177745 0.05844200 0.02993263 0.06544163
#> [18,]  0.02620143 0.016851661 0.09466873 0.09192185 0.02647109 0.03194165
#> [19,]  0.09440438 0.046649581 0.07166570 0.05308175 0.03308526 0.05582359
#> [20,]  0.12295686 0.080381025 0.04893539 0.04421391 0.01883951 0.04708859
#> [21,]  0.06150263 0.014018095 0.04633659 0.03713509 0.02477954 0.07522626
#> [22,]  0.20566446 0.150584222 0.15344473 0.14463463 0.09790248 0.05873741
#> [23,]  0.11870151 0.210068053 0.20975622 0.10579943 0.04348121 0.16676209
#> [24,]  0.08995321 0.074060531 0.04332437 0.03263173 0.01883330 0.01113085
#> [25,]  0.04948636 0.019063829 0.04477098 0.02959630 0.04233120 0.04925125
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.028281634 0.023557492 0.023478882 0.011916665 0.012793204 0.005940700
#>  [2,] 0.021974756 0.024703144 0.014987403 0.009444138 0.011271014 0.004468613
#>  [3,] 0.042141850 0.027632523 0.021272882 0.007609341 0.007254369 0.007445672
#>  [4,] 0.011052074 0.012350386 0.008632778 0.018854888 0.021410058 0.004815037
#>  [5,] 0.028765118 0.014989268 0.003397586 0.010199850 0.015241534 0.014842290
#>  [6,] 0.014391335 0.014845036 0.011039292 0.002554950 0.003302564 0.002680972
#>  [7,] 0.025326572 0.012384681 0.005782711 0.002256079 0.001583649 0.008074784
#>  [8,] 0.006100000 0.025700000 0.008900000 0.026200000 0.010800000 0.013800000
#>  [9,] 0.025905969 0.018250074 0.029215245 0.013803943 0.012644789 0.019419544
#> [10,] 0.022711044 0.017659286 0.012001253 0.005836445 0.010223816 0.003862415
#> [11,] 0.019072335 0.021338765 0.017324999 0.012813587 0.017628110 0.017287423
#> [12,] 0.017433417 0.014985787 0.015955421 0.014911586 0.014178628 0.007011266
#> [13,] 0.025816279 0.026883986 0.018501336 0.007994765 0.009886724 0.011445069
#> [14,] 0.008534568 0.003795268 0.003238834 0.008168598 0.007752578 0.007764408
#> [15,] 0.026429302 0.012794434 0.008562574 0.014601094 0.012484123 0.010392482
#> [16,] 0.021181284 0.010850621 0.009516068 0.007403543 0.007807066 0.008135179
#> [17,] 0.020760957 0.011734780 0.008789975 0.005568131 0.006699011 0.010279458
#> [18,] 0.011970665 0.006634076 0.022834187 0.026637104 0.013547398 0.023686082
#> [19,] 0.020125973 0.012513243 0.010700885 0.010153531 0.005212269 0.004061544
#> [20,] 0.010172712 0.009134186 0.005799919 0.007929260 0.007800265 0.006874771
#> [21,] 0.009309885 0.016078113 0.002891245 0.007861283 0.011233660 0.008256226
#> [22,] 0.024580969 0.018050116 0.010723376 0.005220892 0.013035397 0.007958802
#> [23,] 0.015542996 0.009272042 0.004886566 0.004316127 0.008163449 0.010801089
#> [24,] 0.010636061 0.009349453 0.007813400 0.003671538 0.005636792 0.004457067
#> [25,] 0.017818562 0.005663732 0.007583995 0.004606218 0.019182041 0.019980101
#>                 V56          V57         V58         V59          V6
#>  [1,] -0.0006659181 0.0073116159 0.004454451 0.000435225 0.118507275
#>  [2,]  0.0064479882 0.0034704851 0.009800436 0.007110395 0.087609997
#>  [3,]  0.0079203280 0.0066434256 0.003200640 0.003058496 0.095356412
#>  [4,]  0.0179430753 0.0107792727 0.010396230 0.021164558 0.157889060
#>  [5,]  0.0128539707 0.0095951092 0.012021507 0.010210446 0.068153184
#>  [6,]  0.0029438604 0.0019361265 0.003939207 0.004968909 0.149870916
#>  [7,]  0.0046847704 0.0024255078 0.006923530 0.007203336 0.057703530
#>  [8,]  0.0187000000 0.0230000000 0.005700000 0.011300000 0.119000000
#>  [9,]  0.0112328675 0.0040784932 0.010867103 0.017660142 0.083491786
#> [10,]  0.0063009443 0.0059288716 0.007516547 0.009315845 0.122455388
#> [11,]  0.0085761589 0.0080453730 0.013377893 0.011085718 0.166384585
#> [12,]  0.0117774957 0.0069864384 0.003950419 0.005376554 0.129238439
#> [13,]  0.0064333487 0.0050603934 0.006828021 0.007824836 0.143266121
#> [14,]  0.0034920083 0.0038279235 0.003240797 0.003371229 0.088898308
#> [15,]  0.0084976575 0.0090152491 0.007689792 0.008067752 0.098735960
#> [16,]  0.0083843710 0.0079553973 0.003586486 0.004441081 0.111419498
#> [17,]  0.0062967899 0.0126579738 0.010319649 0.009481739 0.112503988
#> [18,]  0.0031637084 0.0009435008 0.003701202 0.007964471 0.039672714
#> [19,]  0.0093889648 0.0072856883 0.003318492 0.004842667 0.082102763
#> [20,]  0.0075251621 0.0060963427 0.004287136 0.007924507 0.077683194
#> [21,]  0.0056621129 0.0002146413 0.002153358 0.002503698 0.089932230
#> [22,]  0.0052206020 0.0091579486 0.007952324 0.011322103 0.101755654
#> [23,]  0.0067516186 0.0072742203 0.007038729 0.004320968 0.153470202
#> [24,]  0.0052115039 0.0035910676 0.004628869 0.002059883 0.058576692
#> [25,]  0.0045577295 0.0135588448 0.021428348 0.014582695 0.009068981
#>                V60         V7          V8          V9
#>  [1,] 0.0018912204 0.16184865  0.25462156  0.45462061
#>  [2,] 0.0038518893 0.11083988  0.11301540  0.18555495
#>  [3,] 0.0016577103 0.10743642  0.11779622  0.28215344
#>  [4,] 0.0132363916 0.09221693  0.15512520  0.06670494
#>  [5,] 0.0073443628 0.06540045  0.13091715  0.25522617
#>  [6,] 0.0008238459 0.15898570  0.13475679  0.17116503
#>  [7,] 0.0032305282 0.05703128  0.14959623  0.20639683
#>  [8,] 0.0131000000 0.08840000  0.09070000  0.21070000
#>  [9,] 0.0094248609 0.09979282  0.08828590  0.13233918
#> [10,] 0.0087343144 0.14335913  0.16174885  0.22448042
#> [11,] 0.0063831371 0.17581826  0.16975324  0.19378399
#> [12,] 0.0034250750 0.14485490  0.15269483  0.18715020
#> [13,] 0.0059578231 0.13936171  0.20854185  0.29773387
#> [14,] 0.0026829047 0.09808891  0.09739863  0.16771640
#> [15,] 0.0030789521 0.08833992  0.06558023 -0.03153956
#> [16,] 0.0044076484 0.15148258  0.18098576  0.24524521
#> [17,] 0.0082629859 0.14542517  0.13902501  0.11949815
#> [18,] 0.0083926856 0.02892497  0.05532535  0.11288626
#> [19,] 0.0034196559 0.14200546  0.11626593  0.11558704
#> [20,] 0.0065885293 0.09082889  0.08230059  0.08109626
#> [21,] 0.0025424147 0.10171230  0.04813932  0.04483403
#> [22,] 0.0073250098 0.15700557  0.20328258  0.23713955
#> [23,] 0.0051951589 0.08935700  0.08309089  0.17961939
#> [24,] 0.0032191345 0.07510692  0.07952050  0.11375309
#> [25,] 0.0049728453 0.03181035 -0.03020629  0.03866859
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913