Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.011468828 0.17536261 0.21791576 0.24665070 0.40934128 0.61579741
#>  [2,] 0.023070960 0.19794400 0.24181775 0.29298004 0.27669724 0.31621611
#>  [3,] 0.004166126 0.11355354 0.20477318 0.24010109 0.20886492 0.14732323
#>  [4,] 0.052377774 0.31463882 0.24398052 0.19430623 0.38583322 0.48654120
#>  [5,] 0.007061947 0.16264399 0.19294763 0.23434882 0.34393291 0.42346801
#>  [6,] 0.030468352 0.44036237 0.54064112 0.53235337 0.54724747 0.38471284
#>  [7,] 0.042866127 0.26519338 0.38130740 0.49375943 0.52134150 0.55061414
#>  [8,] 0.021026825 0.25931923 0.31250855 0.34485650 0.28307251 0.22826829
#>  [9,] 0.033293979 0.41476233 0.48897853 0.50023203 0.44992317 0.33911867
#> [10,] 0.027041056 0.21056645 0.27089285 0.30411414 0.22070438 0.14792413
#> [11,] 0.057157755 0.36436090 0.33300528 0.27190279 0.27742063 0.28338769
#> [12,] 0.029113219 0.23418402 0.29242724 0.33358023 0.34898359 0.39526152
#> [13,] 0.041219007 0.04183731 0.15164811 0.20869304 0.25267307 0.35109726
#> [14,] 0.039049826 0.24098618 0.30127805 0.33789338 0.30564444 0.26465461
#> [15,] 0.009570133 0.13476099 0.11429281 0.12326529 0.20413289 0.19676825
#> [16,] 0.025931490 0.18579982 0.19296485 0.17788442 0.20696007 0.26751853
#> [17,] 0.023785393 0.14947442 0.16532909 0.15998845 0.23540294 0.36337398
#> [18,] 0.021955104 0.10402684 0.13002733 0.06144245 0.09348087 0.08941341
#> [19,] 0.017492217 0.11973823 0.14722354 0.15707926 0.13437579 0.16450838
#> [20,] 0.035368028 0.34851606 0.37425464 0.48637823 0.61450068 0.74448893
#> [21,] 0.024721641 0.07085796 0.06568318 0.07801075 0.01052409 0.10288863
#> [22,] 0.024377845 0.04987656 0.22806643 0.37412864 0.28029020 0.28354629
#> [23,] 0.010252788 0.17355606 0.14609346 0.10918285 0.23077967 0.31799514
#> [24,] 0.004779348 0.03609068 0.04142240 0.08089883 0.12990711 0.11605787
#> [25,] 0.028176200 0.06262337 0.06458955 0.23788688 0.20223661 0.10251291
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.62110357 0.5566627 0.6346515 0.7373007 0.8290484 0.033878732 0.8722900
#>  [2,] 0.36514353 0.4615963 0.5161178 0.6402581 0.7469598 0.009753298 0.8650546
#>  [3,] 0.02981700 0.1414648 0.2943729 0.4273991 0.4943694 0.015171136 0.6128636
#>  [4,] 0.19844819 0.3647457 0.5390802 0.5776355 0.7856923 0.050466639 0.8449915
#>  [5,] 0.50809758 0.6593385 0.6877095 0.7893856 0.8986850 0.015970394 0.9880482
#>  [6,] 0.28161451 0.3074807 0.3043430 0.2857063 0.5405718 0.055250094 0.7757956
#>  [7,] 0.63235775 0.7815019 0.8864617 0.9230517 0.9426381 0.071216997 0.8919099
#>  [8,] 0.21379730 0.2720871 0.2441872 0.3547647 0.5522067 0.035588848 0.6739570
#>  [9,] 0.24161919 0.2628648 0.3496173 0.4029972 0.5313284 0.029102278 0.6646360
#> [10,] 0.17211942 0.1611834 0.1579812 0.2450173 0.2308094 0.015813768 0.2744034
#> [11,] 0.31403296 0.3309827 0.2647169 0.1987538 0.2764329 0.072124598 0.4726705
#> [12,] 0.45803640 0.5229247 0.6107422 0.6768321 0.7401443 0.035817646 0.7798474
#> [13,] 0.49461991 0.6209376 0.7797143 0.7886315 0.8954259 0.045957193 0.9721454
#> [14,] 0.20881140 0.1474672 0.1967111 0.2477795 0.2683798 0.051207849 0.3236229
#> [15,] 0.20179861 0.3085502 0.3152691 0.3166458 0.4011702 0.013283125 0.4178632
#> [16,] 0.34464971 0.6035400 0.7556218 0.6647788 0.4198937 0.025051107 0.3462832
#> [17,] 0.55951006 0.7388773 0.8413036 0.8475621 0.8630034 0.021507979 0.8474108
#> [18,] 0.11335921 0.1153891 0.1262585 0.1728404 0.2030509 0.016181093 0.2666200
#> [19,] 0.07986600 0.1040339 0.2258471 0.3160183 0.3348329 0.014020690 0.5038564
#> [20,] 0.87260197 0.9137464 0.9078996 0.8575842 0.7036601 0.049745294 0.5849714
#> [21,] 0.07462800 0.1874408 0.1478882 0.2766071 0.3817078 0.034365736 0.3522394
#> [22,] 0.30015334 0.4176342 0.5911802 0.4811810 0.4349174 0.058357222 0.6275170
#> [23,] 0.36461099 0.2460935 0.1033113 0.1675683 0.2165168 0.020217086 0.2444111
#> [24,] 0.07978283 0.1184665 0.1547295 0.1857841 0.1789754 0.006792437 0.1711536
#> [25,] 0.14376324 0.2123794 0.2494344 0.2196428 0.1959063 0.035510899 0.1263569
#>             V21        V22       V23       V24       V25       V26       V27
#>  [1,] 0.9672980 0.99771244 0.9118747 0.9065104 0.9321250 0.8913802 0.7748898
#>  [2,] 0.9047590 0.93662643 0.9241837 0.9388164 0.9883470 0.9306384 0.8618220
#>  [3,] 0.7143195 0.82889513 0.8618150 0.7868351 0.8407905 0.9941575 1.0375653
#>  [4,] 0.5995759 0.69097457 0.8547325 0.7211905 0.7924330 0.8257210 0.7109212
#>  [5,] 1.0581959 1.04624867 0.9559638 0.7858782 0.6039136 0.6265906 0.7527650
#>  [6,] 0.8686390 0.83896185 0.6977612 0.6909726 0.7197440 0.8679501 0.9789923
#>  [7,] 0.7610457 0.59689988 0.4803429 0.3714436 0.2663294 0.1540550 0.1364346
#>  [8,] 0.8063729 0.81302335 0.8241893 0.9120664 0.9026532 0.9024194 0.9170240
#>  [9,] 0.7456770 0.85823447 0.8994923 0.9528503 0.9805569 0.9156630 0.8871152
#> [10,] 0.3980644 0.44570841 0.5611482 0.6978647 0.8022635 0.8489246 0.9235725
#> [11,] 0.4957594 0.37901032 0.3846914 0.4260587 0.4892624 0.5810211 0.6206661
#> [12,] 0.8155158 0.86804512 0.9164133 0.9816117 0.9526138 0.9429437 0.9275059
#> [13,] 0.9248025 0.87743264 0.8110309 0.7447248 0.6123627 0.5416254 0.4975704
#> [14,] 0.4419607 0.52960149 0.6141429 0.6744035 0.7090418 0.8179849 0.8829392
#> [15,] 0.4487549 0.45176518 0.4916886 0.5396178 0.6786631 0.8326958 0.8352345
#> [16,] 0.4551930 0.56005084 0.6416897 0.7245996 0.8369140 0.8886328 0.8974717
#> [17,] 0.9051424 0.92702060 0.8494519 0.8291080 0.8468362 0.8630315 0.7688590
#> [18,] 0.3731844 0.61830122 0.8260580 0.8498354 0.8164866 0.8306376 0.7585005
#> [19,] 0.6284427 0.73423311 0.8438426 0.8663963 0.9174796 0.9490967 0.9005376
#> [20,] 0.3367207 0.43171224 0.5416685 0.3350617 0.2679337 0.4705479 0.5192356
#> [21,] 0.2797869 0.11102580 0.1211403 0.2912046 0.1537131 0.2575074 0.4711380
#> [22,] 0.3651568 0.18699889 0.3666885 0.9686770 0.9874190 0.4742465 0.2809680
#> [23,] 0.2312087 0.31273335 0.5725250 0.7531395 0.7177757 0.7295308 0.7350294
#> [24,] 0.2412671 0.38720134 0.4179909 0.5377564 0.6216995 0.6997822 0.7558292
#> [25,] 0.1794347 0.08358001 0.1862510 0.1925493 0.1052887 0.2043388 0.4372128
#>              V28       V29         V3       V30       V31       V32        V33
#>  [1,] 0.64660639 0.5969014 0.02283356 0.4070098 0.2456327 0.1190570 0.07985589
#>  [2,] 0.75195305 0.6944736 0.01276919 0.5760291 0.3345663 0.2145445 0.20204517
#>  [3,] 0.99829373 0.8850456 0.02270887 0.6454115 0.2598556 0.2342017 0.28608321
#>  [4,] 0.61763868 0.3875866 0.05879822 0.6472035 0.7299828 0.8289450 0.75591755
#>  [5,] 0.79642530 0.6331068 0.01941099 0.4945757 0.3575390 0.2506616 0.07580118
#>  [6,] 0.97764229 0.8789548 0.04177389 0.5836164 0.4668302 0.5047903 0.43210592
#>  [7,] 0.09314854 0.1277771 0.07846718 0.2314033 0.2366454 0.2713831 0.28388293
#>  [8,] 0.94123267 0.8272300 0.03543427 0.5797973 0.3458792 0.2492977 0.20950461
#>  [9,] 0.69843319 0.4107008 0.03066136 0.3457320 0.3522648 0.4068563 0.44476006
#> [10,] 0.98302889 0.8636440 0.04103077 0.6736323 0.6963094 0.6014829 0.42432724
#> [11,] 0.78738766 0.7657765 0.05822380 0.8287731 0.7253548 0.6753983 0.61922633
#> [12,] 0.85691178 0.7230366 0.04230950 0.5681153 0.4316997 0.3048258 0.22718345
#> [13,] 0.35843312 0.3300960 0.07834353 0.3680891 0.2406183 0.2801752 0.40682244
#> [14,] 0.90622416 0.8352191 0.05596784 0.6475287 0.6425997 0.4699060 0.32530838
#> [15,] 0.67287888 0.4587600 0.02921103 0.3575520 0.4027200 0.4163686 0.35258220
#> [16,] 0.83154327 0.6857127 0.03058067 0.7021779 0.7508613 0.6534791 0.57934312
#> [17,] 0.58763026 0.4963074 0.03593537 0.4930329 0.4773991 0.3837965 0.36472065
#> [18,] 0.80629191 0.9118502 0.02585249 0.8214699 0.6840467 0.6212652 0.71374336
#> [19,] 0.79591497 0.7366806 0.02566342 0.6506823 0.5904222 0.5457763 0.53917555
#> [20,] 0.47079102 0.4208057 0.04682778 0.4948232 0.5579518 0.4598384 0.29102470
#> [21,] 0.78916728 0.9512451 0.07486693 0.7913968 0.5010116 0.2419250 0.53341422
#> [22,] 0.66460855 0.8181409 0.06771453 0.4011779 0.5286421 0.7145663 0.56539250
#> [23,] 0.87446619 0.9635226 0.03819724 0.8645533 0.7496610 0.6330750 0.65078049
#> [24,] 0.87016934 0.9813313 0.01466020 0.9187615 0.7572094 0.7218421 0.76824451
#> [25,] 0.67308257 0.7418244 0.02767947 0.6805158 0.7542890 0.6434214 0.65891495
#>              V34        V35        V36       V37        V38       V39
#>  [1,] 0.04564201 0.09093618 0.20433923 0.2044043 0.30514142 0.4057887
#>  [2,] 0.12235457 0.17191341 0.17431065 0.2324764 0.30638915 0.3948602
#>  [3,] 0.43842817 0.73183972 0.89981219 0.8477333 0.66064629 0.4907402
#>  [4,] 0.63898584 0.40016004 0.16591596 0.1083945 0.36065491 0.5780695
#>  [5,] 0.08247951 0.24208671 0.32819291 0.2672038 0.19964703 0.2071379
#>  [6,] 0.29395857 0.09110153 0.24948378 0.1833381 0.38496567 0.6057805
#>  [7,] 0.34413368 0.36383146 0.26207569 0.3578295 0.36282718 0.2015687
#>  [8,] 0.21317840 0.17747636 0.06411856 0.0596631 0.09381196 0.1179060
#>  [9,] 0.41046389 0.23075652 0.34628143 0.4005787 0.33687088 0.2988628
#> [10,] 0.30434600 0.16294234 0.08972009 0.1992660 0.19087181 0.2503039
#> [11,] 0.64178607 0.67988076 0.61586612 0.5350784 0.53953672 0.4678271
#> [12,] 0.16549976 0.15459642 0.10342258 0.1458170 0.21505403 0.1872249
#> [13,] 0.38722035 0.35851458 0.36852675 0.4394820 0.47596887 0.4163458
#> [14,] 0.20436068 0.12015738 0.05329990 0.1784858 0.14925729 0.2284469
#> [15,] 0.33813635 0.43851225 0.44907239 0.3830184 0.30946136 0.3713017
#> [16,] 0.59773002 0.74202409 0.74778502 0.6853251 0.60243338 0.4009537
#> [17,] 0.33634435 0.20890560 0.11184283 0.1916257 0.15999741 0.1366458
#> [18,] 0.82287805 0.76921712 0.57153596 0.3106805 0.19427266 0.1899153
#> [19,] 0.55781066 0.58668914 0.55766382 0.4927305 0.36329339 0.1934731
#> [20,] 0.21577672 0.17444285 0.31997576 0.4132839 0.38826680 0.3747814
#> [21,] 0.61490476 0.31166901 0.41107809 0.6178682 0.61821582 0.5262846
#> [22,] 0.89018474 0.77153930 0.69893992 0.7492673 0.40624089 0.2711645
#> [23,] 0.57244535 0.44046004 0.37339025 0.2207436 0.14944045 0.2074273
#> [24,] 0.60423014 0.50588226 0.46004249 0.4050264 0.40830150 0.3858428
#> [25,] 0.81769322 0.90319879 0.90771462 0.7750071 0.69218477 0.6881971
#>               V4       V40       V41        V42         V43         V44
#>  [1,] 0.03040063 0.3016650 0.1877666 0.16936447  0.13311067  0.17545511
#>  [2,] 0.08880199 0.3952385 0.3530087 0.37695894  0.30184674  0.19142970
#>  [3,] 0.02799948 0.3477711 0.3266728 0.23612958  0.34108621  0.37761686
#>  [4,] 0.09389092 0.4445455 0.4687916 0.48025813  0.39983664  0.33659871
#>  [5,] 0.03366445 0.1788681 0.2239783 0.29578411  0.26335800  0.21422095
#>  [6,] 0.03515707 0.4923107 0.2184934 0.42437609  0.39968481  0.44502693
#>  [7,] 0.07659790 0.2375511 0.4065439 0.34444381  0.17874646  0.11276158
#>  [8,] 0.04901520 0.1280326 0.1055308 0.10948197  0.17438457  0.20987671
#>  [9,] 0.03002308 0.2721881 0.1667442 0.17500964  0.32551889  0.36510853
#> [10,] 0.02468473 0.2274062 0.2931791 0.30117294  0.19473983  0.12179033
#> [11,] 0.07784780 0.4001591 0.4704961 0.47032941  0.41382039  0.41845221
#> [12,] 0.06227471 0.1776132 0.2259399 0.21317675  0.16708453  0.14252857
#> [13,] 0.10616034 0.3440231 0.3308862 0.40526885  0.30975489  0.22053006
#> [14,] 0.08682603 0.2118287 0.2068579 0.27686309  0.18794208  0.08926734
#> [15,] 0.02930525 0.4035359 0.2946882 0.21631803  0.18312504  0.17425375
#> [16,] 0.03288408 0.2755536 0.2300702 0.20828582  0.21931559  0.17492614
#> [17,] 0.03602033 0.1928495 0.1463026 0.11939789  0.08946632  0.11871091
#> [18,] 0.03581316 0.2969627 0.2789884 0.26923822  0.31489836  0.21917902
#> [19,] 0.02567563 0.2253921 0.2344780 0.19734293  0.17602605  0.11574638
#> [20,] 0.03684598 0.3167317 0.2690157 0.23155919  0.10807645  0.13630357
#> [21,] 0.08188496 0.3025735 0.2130414 0.04947828 -0.01425966 -0.02307659
#> [22,] 0.09268021 0.3050551 0.2858870 0.29589215  0.04470507  0.17975826
#> [23,] 0.03562902 0.1376007 0.1274852 0.19696395  0.26226615  0.18115292
#> [24,] 0.02761491 0.3108401 0.1555998 0.05717537  0.09542806  0.13682951
#> [25,] 0.01650864 0.7125948 0.6282501 0.56052116  0.40452985  0.22878004
#>               V45         V46         V47         V48         V49         V5
#>  [1,]  0.14530613  0.11604401  0.13431790  0.11252027  0.08183135 0.11018482
#>  [2,]  0.16459829  0.11599723  0.09384031  0.06522034  0.05287275 0.12705498
#>  [3,]  0.34626800  0.26500177  0.12275282  0.04331616  0.04056457 0.03226208
#>  [4,]  0.44295262  0.46205323  0.27509080  0.21690915  0.12242336 0.19854171
#>  [5,]  0.13286123  0.09528578  0.07740567  0.10787499  0.06953199 0.11107656
#>  [6,]  0.46642614  0.17305730  0.10465929  0.15013546  0.09516341 0.06125746
#>  [7,]  0.07830617  0.06842797  0.07521134  0.05323155  0.03516175 0.10009179
#>  [8,]  0.22351689  0.24655497  0.19222108  0.11120254  0.05936889 0.06605890
#>  [9,]  0.30675031  0.19696617  0.10001671  0.06577049  0.03241202 0.03852996
#> [10,]  0.08460763  0.21950001  0.22233716  0.14051420  0.04393973 0.05142527
#> [11,]  0.52115885  0.44569200  0.26337630  0.18016520  0.09730667 0.07320301
#> [12,]  0.13461419  0.08162155  0.06986198  0.05629955  0.03313266 0.13219515
#> [13,]  0.16385642  0.08403076  0.06093275  0.06074323  0.02918036 0.16530387
#> [14,]  0.10381463  0.17220445  0.19500469  0.14235648  0.06521912 0.09988747
#> [15,]  0.10567705  0.09767677  0.06170168  0.04908717  0.02481168 0.04960039
#> [16,]  0.16772392  0.15359425  0.09724016  0.06151327  0.03580848 0.04354503
#> [17,]  0.13289069  0.10517565  0.10476276  0.07537140  0.03563083 0.08642275
#> [18,]  0.12408855  0.09336723  0.06994618  0.04843558  0.01723751 0.05472201
#> [19,]  0.08036884  0.06839595  0.04220041  0.04300819  0.01835000 0.05832713
#> [20,]  0.11570722  0.08063345  0.06609602  0.05186094  0.02681115 0.05855047
#> [21,] -0.04401880 -0.03412483 -0.02203553 -0.02339681 -0.01180307 0.05440626
#> [22,]  0.13581478  0.15318785  0.12294825  0.07614690  0.04601505 0.16748855
#> [23,]  0.03199794  0.03538305  0.03510324  0.03548427  0.01390978 0.02065085
#> [24,]  0.15035357  0.07278884  0.04063016  0.05860266  0.02738147 0.02640882
#> [25,]  0.15976805  0.11544014  0.13100726  0.11089868  0.06561273 0.03819223
#>                V50          V51         V52         V53         V54         V55
#>  [1,]  0.030860165  0.024437055 0.013964852 0.013075048 0.022431843 0.021294843
#>  [2,]  0.026870560  0.017079157 0.009952787 0.016569584 0.006886202 0.006354968
#>  [3,]  0.010310456  0.002443025 0.011396902 0.006969432 0.003740919 0.002380578
#>  [4,]  0.069610090  0.019990881 0.017354889 0.027024663 0.017870767 0.034693025
#>  [5,]  0.014216749  0.021109769 0.029142063 0.015824605 0.026317041 0.007874170
#>  [6,]  0.018173898  0.033462255 0.014933828 0.005113577 0.008415989 0.009892840
#>  [7,]  0.024174911  0.013804845 0.030173207 0.019934474 0.012908229 0.023538041
#>  [8,]  0.017449668  0.018702045 0.011898466 0.004457448 0.006137906 0.004707912
#>  [9,]  0.013057626  0.011267245 0.006837859 0.006101072 0.010894317 0.008171508
#> [10,]  0.019769460  0.025438645 0.014737645 0.008865940 0.004872260 0.003604787
#> [11,]  0.039084469  0.029950122 0.028033983 0.017773326 0.016688593 0.015120525
#> [12,]  0.017561764  0.016711151 0.019475928 0.016383247 0.015992243 0.005633112
#> [13,]  0.003052685  0.012408062 0.009648824 0.010891395 0.017353035 0.009897091
#> [14,]  0.021354488  0.018895410 0.014327216 0.008288087 0.008597201 0.004788878
#> [15,]  0.013075959  0.011090233 0.009078926 0.007412564 0.013173792 0.008400610
#> [16,]  0.024412918  0.010030362 0.013277816 0.004285457 0.007407921 0.008131917
#> [17,]  0.012851866  0.012956018 0.009183714 0.008579892 0.007390382 0.006176047
#> [18,]  0.011426598  0.009879204 0.007948796 0.008554989 0.008442589 0.004720372
#> [19,]  0.007282553  0.003754902 0.004372030 0.009914666 0.008479970 0.007918764
#> [20,]  0.007734568  0.010452707 0.006938133 0.004516605 0.005193932 0.005208658
#> [21,] -0.004018749 -0.003568863 0.004531414 0.025035378 0.020460764 0.003800137
#> [22,]  0.005398455  0.015275670 0.002070827 0.015597010 0.005974126 0.020757939
#> [23,]  0.015507003  0.017261458 0.014741300 0.013858823 0.004847708 0.006664714
#> [24,]  0.007863200  0.006554944 0.006281695 0.010027541 0.005661507 0.005794734
#> [25,]  0.026808328  0.012940011 0.005873622 0.016341519 0.011452983 0.008184293
#>                V56           V57          V58         V59         V6
#>  [1,] 0.0070338587  0.0076601237 9.986898e-03 0.009618734 0.16833349
#>  [2,] 0.0096412300  0.0056916579 5.496881e-03 0.008257169 0.13441080
#>  [3,] 0.0004837301 -0.0009568759 4.168345e-05 0.002010877 0.05506483
#>  [4,] 0.0296140829  0.0272139245 3.444006e-02 0.019845162 0.12403660
#>  [5,] 0.0031322597  0.0035537829 1.596782e-02 0.010498942 0.21030990
#>  [6,] 0.0042350774  0.0053114386 4.477460e-03 0.003674288 0.12202094
#>  [7,] 0.0144453090  0.0061347531 1.107431e-02 0.014482974 0.08750382
#>  [8,] 0.0054123351  0.0039034200 3.246861e-03 0.002293438 0.08948716
#>  [9,] 0.0090429666  0.0061427149 8.580121e-03 0.008047373 0.09312343
#> [10,] 0.0046174237  0.0035882426 5.913431e-03 0.003065692 0.09243039
#> [11,] 0.0108981215  0.0106226652 1.047818e-02 0.009972895 0.12202647
#> [12,] 0.0112163973  0.0083123532 8.292761e-03 0.005807893 0.16158438
#> [13,] 0.0167120382  0.0092219771 8.382005e-03 0.015190742 0.12212355
#> [14,] 0.0053071181  0.0069550595 7.122792e-03 0.005844433 0.15206686
#> [15,] 0.0068919813  0.0080947950 5.991566e-03 0.008082039 0.06505336
#> [16,] 0.0045130202  0.0108794865 1.019296e-02 0.007324863 0.10312202
#> [17,] 0.0084234589  0.0058251559 3.789382e-03 0.004959704 0.11833447
#> [18,] 0.0069256680  0.0121214016 8.550080e-03 0.006056961 0.07012094
#> [19,] 0.0070421488  0.0042102804 3.703896e-03 0.003783825 0.09949634
#> [20,] 0.0113649953  0.0098364781 4.686020e-03 0.006116647 0.13322406
#> [21,] 0.0130110560  0.0160732213 7.338591e-03 0.013856235 0.10297928
#> [22,] 0.0079039170  0.0156696899 4.395774e-03 0.020815050 0.23747452
#> [23,] 0.0068471243  0.0064627599 5.241485e-03 0.013996462 0.06452771
#> [24,] 0.0080224534  0.0037997881 7.245199e-04 0.005786674 0.03990967
#> [25,] 0.0067052275  0.0062348677 1.053034e-02 0.011370097 0.06225380
#>               V60          V7          V8         V9
#>  [1,] 0.005814232  0.20247177  0.17899286 0.18199042
#>  [2,] 0.006115729  0.16315330  0.10223570 0.13255208
#>  [3,] 0.004233740  0.05645873 -0.01000035 0.04506132
#>  [4,] 0.007993882  0.16214999  0.11698317 0.24467524
#>  [5,] 0.002156269  0.15135414  0.09370706 0.13139535
#>  [6,] 0.005551119  0.14362039  0.19281911 0.28111658
#>  [7,] 0.008290501  0.08077834  0.06315592 0.13375993
#>  [8,] 0.005238544  0.12962649  0.15332191 0.20291363
#>  [9,] 0.005883259  0.12884464  0.20053948 0.36664104
#> [10,] 0.002231012  0.09888972  0.07692442 0.11126758
#> [11,] 0.008135414  0.14687716  0.21963585 0.36157486
#> [12,] 0.004557317  0.15086662  0.18636109 0.22539350
#> [13,] 0.009054525  0.05599196  0.09861170 0.10113408
#> [14,] 0.007329694  0.13754321  0.15532599 0.18901498
#> [15,] 0.004619565  0.07222881  0.05503619 0.08094618
#> [16,] 0.008473398  0.14208787  0.13896599 0.13102421
#> [17,] 0.004854019  0.12042106  0.09540986 0.14253563
#> [18,] 0.006767786  0.08257082  0.08458076 0.06146325
#> [19,] 0.006000152  0.10089470  0.09123605 0.12304466
#> [20,] 0.006406834  0.14750088  0.23452824 0.27993650
#> [21,] 0.011058926 -0.01006191  0.03761414 0.08674684
#> [22,] 0.006816496  0.20968434  0.18563828 0.11780840
#> [23,] 0.008175312  0.06319152  0.11101533 0.15601958
#> [24,] 0.003738818  0.04656849  0.02402201 0.03078369
#> [25,] 0.008129417  0.05938664  0.06508862 0.02532546
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551