Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.067479004 0.41501393 0.35825573 0.14097463 0.08711994 0.21782769
#>  [2,] 0.050741452 0.15251678 0.27119374 0.33291864 0.42504038 0.47528903
#>  [3,] 0.024245631 0.19584972 0.27562341 0.33070638 0.30956331 0.23698080
#>  [4,] 0.020496536 0.11368461 0.15479869 0.21266361 0.23638313 0.35423490
#>  [5,] 0.019014161 0.42926933 0.40473642 0.43054181 0.45402083 0.49479442
#>  [6,] 0.043237535 0.32891918 0.35240360 0.36411448 0.26782319 0.12822945
#>  [7,] 0.036838427 0.23174352 0.39306979 0.50782250 0.55844051 0.55474506
#>  [8,] 0.115270469 0.16807936 0.26803438 0.21827209 0.27879149 0.33075534
#>  [9,] 0.029239941 0.28275546 0.37549302 0.45077882 0.43596166 0.42051898
#> [10,] 0.044252226 0.18094642 0.24314042 0.26572570 0.34100995 0.30189840
#> [11,] 0.028167039 0.28321362 0.32842686 0.34635693 0.34908167 0.35200839
#> [12,] 0.034163312 0.15445353 0.19822385 0.28612781 0.33654347 0.38731143
#> [13,] 0.010010993 0.10263807 0.21801680 0.22778905 0.21734147 0.15305785
#> [14,] 0.029219133 0.23922960 0.28268325 0.32817247 0.28407267 0.23380348
#> [15,] 0.027540202 0.06553996 0.10588810 0.21490047 0.25965473 0.10788395
#> [16,] 0.014081678 0.12282062 0.11058842 0.06834643 0.05320155 0.11781639
#> [17,] 0.016317289 0.05561195 0.11951382 0.10749870 0.12234433 0.12672707
#> [18,] 0.020515880 0.05805970 0.05644803 0.15841116 0.13319426 0.17104991
#> [19,] 0.026917857 0.07000559 0.20128326 0.25730435 0.30006175 0.26953729
#> [20,] 0.037307764 0.29446472 0.19848049 0.10172216 0.13343299 0.42616887
#> [21,] 0.012246501 0.11672341 0.06311934 0.04011401 0.09317742 0.11404061
#> [22,] 0.023989750 0.13499355 0.16632098 0.17555507 0.17647441 0.15932048
#> [23,] 0.020215782 0.09255318 0.06825966 0.07701088 0.10949114 0.02914671
#> [24,] 0.007112407 0.09841532 0.09047835 0.08288627 0.09524286 0.09915428
#> [25,] 0.017402017 0.18354967 0.17776048 0.14520348 0.27771308 0.37458471
#>                V15         V16        V17       V18       V19           V2
#>  [1,]  0.335517772 0.440675414 0.32553221 0.2722174 0.3949084 0.0922563099
#>  [2,]  0.596441784 0.555736181 0.68813166 0.8194983 0.9379291 0.0501725028
#>  [3,]  0.196784252 0.148387213 0.12891819 0.2207227 0.2593598 0.0401457415
#>  [4,]  0.504357546 0.594159163 0.66701950 0.7823491 0.8859950 0.0268528551
#>  [5,]  0.460575878 0.599507829 0.80711226 0.8141246 0.8176646 0.1144416049
#>  [6,]  0.100935473 0.203954956 0.18859148 0.2045540 0.2625953 0.0256394564
#>  [7,]  0.616680214 0.745967591 0.90572224 0.9644726 0.9469055 0.0547861763
#>  [8,]  0.374920958 0.413752196 0.41830737 0.4379628 0.5190723 0.2123687848
#>  [9,]  0.409884936 0.449687630 0.53468914 0.6336838 0.6809637 0.0249922628
#> [10,]  0.225505344 0.259186573 0.31133331 0.2106648 0.2577980 0.0645605197
#> [11,]  0.286131219 0.231374695 0.21510524 0.2442064 0.3690509 0.0350330541
#> [12,]  0.418312637 0.528587345 0.64004785 0.7921955 0.8476680 0.0277865360
#> [13,]  0.056744673 0.102268036 0.22997236 0.3603185 0.4428209 0.0156361612
#> [14,]  0.167806803 0.203516374 0.24002519 0.3530779 0.5668106 0.0384682677
#> [15,]  0.066513120 0.137642269 0.20830737 0.3378548 0.1308538 0.0149359770
#> [16,]  0.089548286 0.094693597 0.22232419 0.2825852 0.3309241 0.0038995471
#> [17,]  0.089186880 0.132621652 0.20841615 0.3073322 0.3198353 0.0085043877
#> [18,]  0.197633401 0.207349668 0.20056278 0.1672466 0.1998415 0.0234264554
#> [19,]  0.369995347 0.471859341 0.37972837 0.4594079 0.6774164 0.0288685650
#> [20,]  0.580825476 0.618954213 0.63721433 0.5395779 0.3533056 0.0745444019
#> [21,]  0.121469563 0.223117545 0.24723828 0.2783799 0.4367011 0.0083231338
#> [22,]  0.222705894 0.555201019 0.72132639 0.6201715 0.3790774 0.0121540599
#> [23,] -0.009041619 0.003325045 0.05171367 0.2134037 0.3596094 0.0004295296
#> [24,]  0.153725650 0.157952707 0.12288800 0.1376297 0.1211556 0.0131326460
#> [25,]  0.403177129 0.251376627 0.12982892 0.1955592 0.2383409 0.0248751575
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.7588919 0.8733404 0.6242936 0.4677938 0.4620606 0.3105282 0.4562577
#>  [2,] 0.9133994 1.0021577 1.0240431 0.9591535 0.8147954 0.6654000 0.5089972
#>  [3,] 0.2694485 0.3258315 0.3780906 0.5525203 0.7184981 0.7677107 0.7820802
#>  [4,] 0.9806543 0.9259981 0.8996876 0.9529866 0.8147595 0.4580646 0.3335617
#>  [5,] 0.8509148 0.7821804 0.8421890 1.0568299 1.1278774 1.0137194 0.9881517
#>  [6,] 0.3672381 0.4799964 0.6315082 0.7481606 0.7787801 0.8881421 1.0120964
#>  [7,] 0.8961800 0.8014702 0.6034577 0.4185520 0.3772434 0.3089936 0.1255161
#>  [8,] 0.3460348 0.2110974 0.1620675 0.3888823 0.6374025 0.4990693 0.3077981
#>  [9,] 0.7520170 0.8179437 0.9108095 0.9209565 0.9709509 0.9938278 0.9499074
#> [10,] 0.3047695 0.3786359 0.5040758 0.5787085 0.5890419 0.5614438 0.5887643
#> [11,] 0.5091064 0.5429843 0.4492154 0.4396076 0.4803037 0.5251497 0.7517376
#> [12,] 0.8205036 0.9113897 0.9358512 0.9093784 0.9414297 0.8467036 0.6652568
#> [13,] 0.5595251 0.6701423 0.8056335 0.8323344 0.7571263 0.8222216 0.9612900
#> [14,] 0.6985699 0.8151340 0.8082910 0.8222345 0.9628191 0.9768995 0.9921263
#> [15,] 0.2008737 0.4878994 0.5553793 0.6727240 0.7227476 0.6899743 0.5543380
#> [16,] 0.4951154 0.6244027 0.7543672 0.8913971 0.8852558 0.9161794 0.9989836
#> [17,] 0.3997385 0.5473642 0.5991223 0.6658373 0.7445230 0.8261886 0.9128563
#> [18,] 0.2725103 0.3417485 0.3403668 0.3672953 0.3278480 0.2668894 0.3506092
#> [19,] 0.7223029 0.5616808 0.4315088 0.6259609 0.7408819 0.6699608 0.7541503
#> [20,] 0.3964890 0.6523312 0.8451593 0.8851333 0.7093699 0.6455234 0.6411761
#> [21,] 0.6341254 0.7007545 0.8757511 0.9135900 0.7386122 0.7267589 0.8881204
#> [22,] 0.3571488 0.4631066 0.5610559 0.5964124 0.6547776 0.8370385 0.9799609
#> [23,] 0.4064662 0.4088576 0.4958666 0.6444973 0.7625501 0.8780522 0.8803262
#> [24,] 0.1415359 0.2397819 0.3167387 0.4587324 0.6648899 0.8543697 0.9217723
#> [25,] 0.2886158 0.2978318 0.3470550 0.6141116 0.7944308 0.7450618 0.7141999
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.3007666 0.3614378 0.4834965 0.078114481 0.5521720 0.1841118 0.1913243
#>  [2,] 0.3495198 0.2585548 0.3326525 0.077350045 0.3753574 0.3039470 0.1663358
#>  [3,] 0.8559433 0.9658183 0.9375293 0.060370815 0.6626666 0.6233798 0.4974777
#>  [4,] 0.3143109 0.3999530 0.4797780 0.022681794 0.3951900 0.2873568 0.3794120
#>  [5,] 1.0111570 0.7601994 0.6042032 0.129520595 0.5166689 0.2609889 0.1854326
#>  [6,] 1.0069501 0.9407154 0.8985857 0.033526636 0.8521262 0.7050246 0.3915210
#>  [7,] 0.1134111 0.1325571 0.1099589 0.071810022 0.1438494 0.2316103 0.3213760
#>  [8,] 0.4145883 0.5454271 0.4942335 0.281036942 0.4146930 0.3625944 0.2406185
#>  [9,] 0.9405601 0.8897153 0.7545675 0.028637670 0.5613851 0.3949307 0.3061342
#> [10,] 0.6503532 0.8038890 0.8426395 0.061355340 0.8324657 0.8931020 0.8514789
#> [11,] 0.8858001 0.8858969 0.8723345 0.042968114 0.7600686 0.5938504 0.5441212
#> [12,] 0.5270891 0.3196824 0.2858400 0.031422320 0.3812029 0.3122774 0.2013666
#> [13,] 1.0085986 0.9476338 0.8073981 0.017223491 0.5742004 0.2203854 0.2377299
#> [14,] 0.9974923 0.9982732 0.9026788 0.046696242 0.6705586 0.4209869 0.3043303
#> [15,] 0.4651937 0.4206587 0.4671471 0.016477702 0.4471719 0.4607500 0.5826461
#> [16,] 0.9949413 0.9031580 0.8162452 0.010046881 0.7052698 0.6080657 0.5567173
#> [17,] 0.8488953 0.7090344 0.6680928 0.025062284 0.6073116 0.5827677 0.5652232
#> [18,] 0.4768824 0.5973587 0.6714922 0.022669060 0.7678183 0.6986958 0.5281487
#> [19,] 0.8146004 0.8817440 0.8169734 0.038823356 0.7225338 0.5460007 0.3282741
#> [20,] 0.5701804 0.5255660 0.2746194 0.101729418 0.3095698 0.5105460 0.4886756
#> [21,] 0.9574943 0.9039614 0.7937540 0.001919696 0.6574207 0.5645742 0.4645629
#> [22,] 0.9832118 0.9244779 0.7699410 0.015979357 0.7747522 0.8151598 0.7170265
#> [23,] 0.7758223 0.8245686 0.9681908 0.011869861 0.9653449 0.9253066 0.8966109
#> [24,] 0.7866243 0.6296911 0.5205389 0.005822700 0.3683893 0.2873297 0.1708845
#> [25,] 0.7066773 0.8519639 0.9749332 0.040850042 0.8974239 0.7655538 0.6399215
#>              V33       V34        V35         V36        V37        V38
#>  [1,] 0.42735383 0.5891873 0.76542438  0.71452496 0.59644762 0.66536676
#>  [2,] 0.30441308 0.5519632 0.62190254  0.54913198 0.43623955 0.33830740
#>  [3,] 0.41614482 0.3218454 0.15770813  0.12987919 0.25797501 0.14780227
#>  [4,] 0.31932343 0.1889411 0.20204741  0.28515887 0.31320653 0.38125894
#>  [5,] 0.06350185 0.2900385 0.42908438  0.29455756 0.25573925 0.29895050
#>  [6,] 0.37264695 0.3040290 0.14188208 -0.01872882 0.09468972 0.15033706
#>  [7,] 0.32858490 0.4024014 0.44216704  0.24301426 0.21199412 0.30647164
#>  [8,] 0.13990348 0.1946838 0.28450471  0.40972158 0.31489104 0.31501404
#>  [9,] 0.22540644 0.0294454 0.03292653  0.14814654 0.23944389 0.30150499
#> [10,] 0.71199079 0.6562625 0.67028623  0.48416523 0.29029602 0.33694497
#> [11,] 0.54019757 0.3837620 0.17725167  0.19111364 0.24232656 0.25268165
#> [12,] 0.42561685 0.4808246 0.43667826  0.39525320 0.27046265 0.33358371
#> [13,] 0.33895674 0.4888382 0.74471732  0.92525334 0.87679591 0.66960982
#> [14,] 0.25585838 0.2129830 0.15039628  0.02752713 0.04726530 0.09262726
#> [15,] 0.75412554 0.6199237 0.55592928  0.58286798 0.61186990 0.52050160
#> [16,] 0.52120708 0.5465581 0.56692835  0.53295887 0.47408142 0.36968365
#> [17,] 0.53230353 0.5579151 0.60329734  0.62694738 0.57708574 0.46567468
#> [18,] 0.65380452 0.8171298 0.87401834  0.79743559 0.60353680 0.46544581
#> [19,] 0.34145477 0.1527851 0.24885954  0.51789311 0.46756767 0.14851388
#> [20,] 0.32288605 0.1867220 0.34348114  0.61230454 0.73165183 0.50006819
#> [21,] 0.43343670 0.4766864 0.53290519  0.47743000 0.48504829 0.37046784
#> [22,] 0.68570667 0.6630903 0.69582104  0.66621159 0.63280147 0.57670286
#> [23,] 0.99326043 0.9922086 0.80195170  0.51785633 0.24022304 0.16834671
#> [24,] 0.19354772 0.2284449 0.24085328  0.19515269 0.11143927 0.10133334
#> [25,] 0.68753690 0.5960950 0.43898010  0.36375131 0.23219985 0.15419404
#>             V39          V4        V40        V41        V42         V43
#>  [1,] 0.5883209 0.076064287 0.45233222 0.34127737 0.36243552  0.36850584
#>  [2,] 0.3289237 0.111907750 0.21629112 0.16181713 0.27317757  0.33036558
#>  [3,] 0.2328614 0.061255459 0.28765828 0.28102010 0.33596533  0.18165271
#>  [4,] 0.2810587 0.060630792 0.16512839 0.13460258 0.19175375  0.23321791
#>  [5,] 0.2392162 0.051812554 0.20678976 0.39279568 0.33807855 -0.02088246
#>  [6,] 0.1058199 0.037590805 0.06905162 0.09686853 0.17660434  0.24309710
#>  [7,] 0.2515096 0.089079481 0.12492754 0.12856192 0.17206855  0.16719081
#>  [8,] 0.3782819 0.394269097 0.42383714 0.36112179 0.27890294  0.18057915
#>  [9,] 0.2662925 0.075957815 0.25076948 0.30594024 0.31261244  0.28720146
#> [10,] 0.3061951 0.060228163 0.33979817 0.60147306 0.65463978  0.48206156
#> [11,] 0.3022057 0.054251149 0.33096606 0.38021222 0.46541452  0.43223492
#> [12,] 0.4044423 0.062038556 0.42617177 0.45340757 0.43855942  0.32917447
#> [13,] 0.4811598 0.034875038 0.33344663 0.32197921 0.28738493  0.39751829
#> [14,] 0.0816688 0.035046388 0.10603442 0.02613127 0.10657256  0.22794548
#> [15,] 0.4839731 0.011075214 0.66890701 0.77685107 0.71863924  0.66639074
#> [16,] 0.1720277 0.030800876 0.16426534 0.17515966 0.07228075  0.05071348
#> [17,] 0.3298056 0.019119622 0.33908211 0.29061970 0.23120842  0.19216694
#> [18,] 0.3315832 0.018697347 0.40506455 0.44585562 0.37307984  0.22891848
#> [19,] 0.1661732 0.061098937 0.09410526 0.28824653 0.33395125  0.23126080
#> [20,] 0.3964174 0.104744760 0.49253416 0.50485252 0.35641281  0.24063204
#> [21,] 0.2152463 0.025498107 0.09975889 0.09894829 0.04418231  0.08239655
#> [22,] 0.3459527 0.027114667 0.19127668 0.19983586 0.18526199  0.18513223
#> [23,] 0.1137526 0.001990313 0.29651962 0.37046496 0.40923202  0.32618790
#> [24,] 0.1810126 0.012585293 0.22663577 0.27613904 0.27392387  0.21205902
#> [25,] 0.2458210 0.039350212 0.18378586 0.14300593 0.21578435  0.26418764
#>              V44          V45        V46        V47        V48         V49
#>  [1,] 0.29749415  0.346879245 0.26989288 0.20957554 0.20978960 0.102682755
#>  [2,] 0.22073531  0.112105009 0.07637358 0.08323577 0.01352036 0.034846431
#>  [3,] 0.06356309  0.091192793 0.20552584 0.24002412 0.17258399 0.068158191
#>  [4,] 0.18864123  0.083003803 0.03174096 0.01957863 0.06112162 0.037469149
#>  [5,] 0.13120711  0.187679356 0.16725812 0.08880381 0.06153688 0.008638893
#>  [6,] 0.16389871  0.096421494 0.09912208 0.11215960 0.09032138 0.063391948
#>  [7,] 0.06617023  0.070832173 0.04718739 0.05294539 0.05161216 0.031952652
#>  [8,] 0.20399278  0.285008519 0.22985508 0.18518786 0.15326336 0.069713785
#>  [9,] 0.23320361  0.157880271 0.08199989 0.08509720 0.08315174 0.053485947
#> [10,] 0.47420593  0.564455450 0.51988130 0.31261470 0.17202776 0.075767465
#> [11,] 0.30070678  0.311252523 0.23208847 0.18947227 0.16167396 0.095023259
#> [12,] 0.26074324  0.182875510 0.05686038 0.09920906 0.06177930 0.019248225
#> [13,] 0.41780347  0.394786636 0.28279136 0.13681253 0.04605701 0.041408554
#> [14,] 0.26517420  0.231125575 0.23929088 0.18164693 0.10730515 0.058874727
#> [15,] 0.44442967  0.345533038 0.29761656 0.19635863 0.13522792 0.047107451
#> [16,] 0.03083041  0.048707887 0.02657291 0.01560092 0.01746281 0.007287864
#> [17,] 0.16334251  0.088807943 0.07387667 0.03534650 0.02507917 0.011046363
#> [18,] 0.11116989  0.091656549 0.10014224 0.09179320 0.04452701 0.021580779
#> [19,] 0.10476105  0.122788844 0.14364900 0.06995945 0.06437840 0.022559685
#> [20,] 0.18770375  0.160863782 0.10720152 0.13137434 0.12651175 0.074218494
#> [21,] 0.02313011 -0.002587051 0.03159357 0.04167474 0.03690497 0.019083714
#> [22,] 0.13074385  0.115567515 0.13931826 0.07713163 0.04217816 0.024543435
#> [23,] 0.28253539  0.260048186 0.19119216 0.09242007 0.03347604 0.028173200
#> [24,] 0.14532570  0.116913665 0.11954919 0.05614552 0.04093840 0.031340900
#> [25,] 0.16544465  0.020404609 0.02589573 0.02819409 0.03528828 0.010960975
#>                V5         V50         V51          V52         V53         V54
#>  [1,] 0.026839877 0.031396145 0.025846425 1.410027e-02 0.017630046 0.018945715
#>  [2,] 0.154502159 0.012615671 0.013986688 1.248284e-02 0.003264492 0.026847947
#>  [3,] 0.095607304 0.009310767 0.027213202 2.138254e-02 0.012406639 0.009564005
#>  [4,] 0.108689544 0.007709402 0.013032905 2.224030e-02 0.010233634 0.009363781
#>  [5,] 0.143816473 0.015952659 0.017676333 2.197001e-02 0.025923992 0.035174094
#>  [6,] 0.039756265 0.015903583 0.012837925 1.099132e-02 0.007290962 0.006018158
#>  [7,] 0.103428529 0.018558939 0.020158495 2.036694e-02 0.018290209 0.017918836
#>  [8,] 0.366774232 0.019793404 0.014206209 3.580697e-02 0.021792131 0.015923756
#>  [9,] 0.122659363 0.022277769 0.017756097 1.955599e-02 0.015954167 0.016769606
#> [10,] 0.079411920 0.035280359 0.040334729 3.621334e-02 0.020576656 0.016861596
#> [11,] 0.089075497 0.026370832 0.019334931 1.291918e-02 0.007721061 0.006563515
#> [12,] 0.034131013 0.010413723 0.022285477 6.764587e-03 0.016880794 0.013165434
#> [13,] 0.029289416 0.009718530 0.002862633 1.084942e-02 0.005579155 0.001971884
#> [14,] 0.062301622 0.017379715 0.021768538 1.481227e-02 0.003501062 0.001968272
#> [15,] 0.017352809 0.031023683 0.010981352 7.687934e-05 0.006228972 0.004088503
#> [16,] 0.046676926 0.008262284 0.006132585 5.884510e-03 0.006557099 0.002216045
#> [17,] 0.055084251 0.004904314 0.002497048 4.063919e-03 0.005493571 0.008591172
#> [18,] 0.036415453 0.019577988 0.012703000 8.427935e-03 0.010021727 0.014138593
#> [19,] 0.027297750 0.018769899 0.007075253 1.378288e-02 0.011946680 0.010300188
#> [20,] 0.124781980 0.021490520 0.020153507 2.112224e-02 0.018965824 0.011308741
#> [21,] 0.054993398 0.011739665 0.003230575 5.516052e-03 0.013248444 0.011318476
#> [22,] 0.036106592 0.018024400 0.008778471 1.020577e-02 0.001604537 0.002099794
#> [23,] 0.007433393 0.010644857 0.006461936 1.085458e-02 0.006948144 0.006387913
#> [24,] 0.022285028 0.008559070 0.003532014 7.434485e-03 0.003642648 0.004887819
#> [25,] 0.017317660 0.015016363 0.016816482 1.375042e-02 0.012432509 0.006891431
#>               V55           V56          V57          V58           V59
#>  [1,] 0.009371093  0.0161689187 0.0084498498 0.0122417119  0.0172466681
#>  [2,] 0.013518643  0.0089582873 0.0233239735 0.0366417429  0.0163505973
#>  [3,] 0.006232817  0.0076053903 0.0041054272 0.0055455810  0.0053234974
#>  [4,] 0.016957399  0.0154290300 0.0049357107 0.0086464417  0.0113188593
#>  [5,] 0.016829518  0.0145848628 0.0140761639 0.0116525281 -0.0006757373
#>  [6,] 0.006326260  0.0049239670 0.0074149785 0.0076381208  0.0058371693
#>  [7,] 0.021765222  0.0125991409 0.0062755700 0.0149838672  0.0118372488
#>  [8,] 0.020633160  0.0027141322 0.0134134611 0.0151672896  0.0186203530
#>  [9,] 0.004216320  0.0123859113 0.0075472464 0.0092078115  0.0066723728
#> [10,] 0.014302874  0.0101922894 0.0084689387 0.0104078023  0.0116822190
#> [11,] 0.005546272  0.0055326434 0.0031258363 0.0066633310  0.0076654890
#> [12,] 0.001860132  0.0133951620 0.0093344883 0.0060645050  0.0199760184
#> [13,] 0.001570923 -0.0004833628 0.0006155141 0.0008468691  0.0018723674
#> [14,] 0.003483855  0.0020410180 0.0031106581 0.0016925901  0.0018812617
#> [15,] 0.001021303  0.0093672755 0.0086673028 0.0056124896  0.0079510834
#> [16,] 0.002031608  0.0046144074 0.0028653982 0.0021502850  0.0033852440
#> [17,] 0.011459848  0.0104195345 0.0052795813 0.0010842588  0.0009445764
#> [18,] 0.011018430  0.0065295707 0.0105194724 0.0079271040  0.0117920812
#> [19,] 0.013063038  0.0095059124 0.0088635132 0.0052707363  0.0063439546
#> [20,] 0.006398133  0.0095886471 0.0118858695 0.0110721321  0.0053351424
#> [21,] 0.008400037  0.0030405064 0.0029636298 0.0018815352  0.0005439687
#> [22,] 0.007286264  0.0050311691 0.0114584037 0.0135120519  0.0071798998
#> [23,] 0.002758091  0.0004958231 0.0003067662 0.0021901365  0.0058880531
#> [24,] 0.004867658  0.0028164851 0.0026948475 0.0041613577  0.0016440359
#> [25,] 0.007789318  0.0082235980 0.0060688750 0.0061175576  0.0153480191
#>                V6          V60         V7          V8          V9
#>  [1,] 0.032599237 0.0128740429 0.08728988 0.259502405 0.374498021
#>  [2,] 0.168794790 0.0052245558 0.16639922 0.078910662 0.106160586
#>  [3,] 0.150160710 0.0000278359 0.14280826 0.110144087 0.125159958
#>  [4,] 0.103335490 0.0044743294 0.09272510 0.081962134 0.093028038
#>  [5,] 0.192808544 0.0023585431 0.18138348 0.262512740 0.395631881
#>  [6,] 0.095596598 0.0058206546 0.13844114 0.196829756 0.270447851
#>  [7,] 0.086036125 0.0080857787 0.11140620 0.100018673 0.120965518
#>  [8,] 0.171610852 0.0217438023 0.17860149 0.034491501 0.165708059
#>  [9,] 0.166375196 0.0050808813 0.16881954 0.203533263 0.256750945
#> [10,] 0.154772765 0.0075067906 0.12535591 0.154035036 0.232018387
#> [11,] 0.095876538 0.0036346836 0.10495641 0.135312343 0.238930758
#> [12,] 0.111790682 0.0171667833 0.17665886 0.150514538 0.091642037
#> [13,] 0.053428612 0.0040449946 0.04868162 0.006945608 0.006869519
#> [14,] 0.051594130 0.0027712865 0.10506867 0.126350533 0.184466463
#> [15,] 0.062129840 0.0077694628 0.11167134 0.104500607 0.045978917
#> [16,] 0.079438681 0.0027230921 0.08836116 0.071666900 0.144844483
#> [17,] 0.100165539 0.0040057312 0.10402158 0.074252515 0.071886499
#> [18,] 0.069916280 0.0049889319 0.07129376 0.029721920 0.004054834
#> [19,] 0.046612924 0.0030572439 0.09394997 0.086352277 0.098605942
#> [20,] 0.128513565 0.0076395637 0.10678715 0.164726124 0.209894604
#> [21,] 0.085123591 0.0023591825 0.08017054 0.049292735 0.103997702
#> [22,] 0.081091840 0.0097642302 0.10871372 0.099611054 0.060673422
#> [23,] 0.006206637 0.0044266345 0.04666402 0.009323510 0.031467747
#> [24,] 0.036293642 0.0030764040 0.05956641 0.067049954 0.093173315
#> [25,] 0.065737089 0.0086481431 0.07052609 0.113481643 0.159281496
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203