Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.038598366 0.24272611 0.23266512 0.26438514 0.33048272 0.31896043
#>  [2,] 0.034939554 0.24915847 0.31286419 0.31699771 0.25663684 0.11573610
#>  [3,] 0.048449301 0.54019748 0.40459626 0.14069875 0.25873557 0.34851437
#>  [4,] 0.022913255 0.08105464 0.17551622 0.21181820 0.18944651 0.13583431
#>  [5,] 0.049267255 0.17932495 0.26798534 0.32306787 0.37887347 0.41287481
#>  [6,] 0.030145058 0.34274000 0.40611025 0.40607683 0.38474851 0.35446328
#>  [7,] 0.016985140 0.18536905 0.26941810 0.31136505 0.28279586 0.20558624
#>  [8,] 0.035912430 0.28265076 0.31955482 0.33841924 0.30314635 0.32753338
#>  [9,] 0.029970401 0.20949327 0.24651876 0.29662318 0.27484316 0.24640390
#> [10,] 0.044801389 0.12034514 0.21894491 0.25399887 0.22607166 0.23989368
#> [11,] 0.010050063 0.24980965 0.27797444 0.35397217 0.32378459 0.24055580
#> [12,] 0.026547742 0.26433440 0.21350610 0.13734749 0.27480789 0.42531182
#> [13,] 0.110478467 0.52903021 0.37345130 0.11671410 0.09787164 0.21159792
#> [14,] 0.018876677 0.18641149 0.24586484 0.30093461 0.28997176 0.26183098
#> [15,] 0.026605139 0.00796803 0.12448558 0.31602608 0.30104693 0.26230584
#> [16,] 0.016224631 0.09255186 0.13980008 0.17816285 0.25915246 0.28962291
#> [17,] 0.024335831 0.06063811 0.04840559 0.08693741 0.17457078 0.32566230
#> [18,] 0.014314338 0.15618212 0.15396058 0.07401748 0.09467339 0.10393020
#> [19,] 0.016534776 0.08530489 0.07741449 0.08107659 0.11570438 0.07055830
#> [20,] 0.011676500 0.05585158 0.10532062 0.14336067 0.22229952 0.29386978
#> [21,] 0.049720168 0.13759932 0.19282615 0.28411528 0.27602144 0.18623117
#> [22,] 0.011720693 0.11567481 0.14131561 0.09279590 0.07258308 0.08287993
#> [23,] 0.017978050 0.11521272 0.12858030 0.11342846 0.08922735 0.13414432
#> [24,] 0.007890373 0.18835854 0.12434286 0.16022425 0.33234763 0.23022916
#> [25,] 0.034197473 0.29526852 0.25555356 0.15473508 0.23857799 0.49207892
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.25839963 0.32846988 0.3317350 0.2497974 0.3471718 0.060930098 0.4241733
#>  [2,] 0.11302708 0.15811169 0.1902928 0.2478079 0.2266771 0.023290186 0.2514237
#>  [3,] 0.46516518 0.46766745 0.1381630 0.1157223 0.3189896 0.081897989 0.9040323
#>  [4,] 0.07898651 0.13910850 0.1818369 0.3021285 0.3465768 0.051681306 0.5499480
#>  [5,] 0.49340306 0.57572800 0.7279880 0.7845554 0.8402668 0.059338423 0.8635064
#>  [6,] 0.35556159 0.40350175 0.4371933 0.4837266 0.5943595 0.034110162 0.7074502
#>  [7,] 0.07103132 0.24377864 0.4103962 0.6062081 0.7440281 0.004086898 0.7416946
#>  [8,] 0.26637037 0.18477308 0.2100245 0.2433835 0.3980616 0.039937164 0.5109404
#>  [9,] 0.17714430 0.19871402 0.2196998 0.3161728 0.5230058 0.035012016 0.6595454
#> [10,] 0.20464329 0.22300683 0.2948975 0.3067862 0.3866506 0.066903770 0.3339463
#> [11,] 0.17080805 0.21594051 0.2454549 0.3789090 0.6190880 0.011565382 0.7203456
#> [12,] 0.57029788 0.71232100 0.7077176 0.7860761 0.9202957 0.038424681 0.9637595
#> [13,] 0.39857542 0.40268254 0.2890696 0.2292851 0.3525366 0.107803493 0.8376616
#> [14,] 0.24732180 0.14671526 0.1432188 0.2171652 0.3018761 0.023787724 0.3745823
#> [15,] 0.10960745 0.16963231 0.3146350 0.4603867 0.3792505 0.021294503 0.5579219
#> [16,] 0.30337503 0.35096222 0.3290138 0.3641750 0.4490537 0.021238208 0.5070403
#> [17,] 0.49971533 0.69711801 0.7782470 0.7869800 0.8050538 0.026873362 0.8592994
#> [18,] 0.18970240 0.14037852 0.1436307 0.2221238 0.3164504 0.016314510 0.3695864
#> [19,] 0.04795691 0.07605045 0.1223292 0.2305995 0.3480411 0.005303998 0.3691229
#> [20,] 0.39708679 0.47077290 0.5707986 0.6794300 0.7486835 0.013369791 0.7841357
#> [21,] 0.23190624 0.31523838 0.4053827 0.4497287 0.4444178 0.054871911 0.3529321
#> [22,] 0.12846162 0.11919138 0.1370386 0.1503024 0.2794505 0.008234783 0.3262495
#> [23,] 0.10514836 0.14873638 0.3137769 0.3489142 0.3317603 0.010393962 0.4197076
#> [24,] 0.05208194 0.20212889 0.2823137 0.3378955 0.6114151 0.042098537 0.7178141
#> [25,] 0.64966457 0.73015698 0.7664273 0.6577433 0.4447467 0.049820320 0.3679104
#>             V21       V22        V23       V24       V25       V26       V27
#>  [1,] 0.4451569 0.4866873 0.52651540 0.5700165 0.5784218 0.6539162 0.7405610
#>  [2,] 0.3618528 0.6138203 0.79525178 0.8612462 0.9561814 0.9864672 0.9580096
#>  [3,] 0.9246278 0.3572414 0.18025726 0.1769421 0.1443076 0.3567287 0.2848192
#>  [4,] 0.6158185 0.7139094 0.76529764 0.6228922 0.6559925 0.8158235 0.9668780
#>  [5,] 0.8748528 0.7988754 0.67640353 0.5906205 0.4937449 0.4272495 0.3769171
#>  [6,] 0.8028621 0.8861837 0.89718004 0.9212355 0.8851902 0.8703856 0.8679300
#>  [7,] 0.7892624 0.8450380 0.93376344 0.9412106 0.9791577 0.8210531 0.4882435
#>  [8,] 0.6953743 0.7279502 0.67390364 0.6810847 0.7214079 0.6792683 0.7880635
#>  [9,] 0.7784188 0.7943193 0.77258331 0.8911895 0.9314746 0.9858734 1.0075122
#> [10,] 0.3550282 0.3476278 0.51031355 0.6871835 0.7018358 0.6964666 0.7996170
#> [11,] 0.8236730 0.8102609 0.84249515 0.9499562 0.9488758 0.9617482 0.9797458
#> [12,] 0.9863809 0.9757716 0.95057044 0.7239401 0.4848740 0.4622448 0.5367498
#> [13,] 0.9832401 0.4529945 0.06969526 0.2089586 0.1614075 0.3086536 0.1934004
#> [14,] 0.4619950 0.4732797 0.57746318 0.6684212 0.7852978 0.8226169 0.9336267
#> [15,] 0.6050903 0.5896741 0.68683434 0.8536706 0.9447402 0.7400168 0.5719384
#> [16,] 0.5488372 0.5785251 0.59434817 0.7162366 0.8501989 0.8795749 0.9030393
#> [17,] 0.9848554 0.9243762 0.87891504 0.9249515 0.9218556 0.9441897 0.9011807
#> [18,] 0.4209773 0.7670502 0.84580668 0.7483903 0.6179208 0.6862224 0.6346348
#> [19,] 0.3667637 0.4349697 0.52959495 0.6480628 0.7249757 0.7443278 0.7242183
#> [20,] 0.8931021 0.9418249 0.93779733 0.9079158 0.8318611 0.7619412 0.6203194
#> [21,] 0.3041542 0.2289527 0.21046029 0.2381158 0.2921147 0.3671029 0.4157186
#> [22,] 0.2823119 0.4001840 0.74295154 0.8603645 0.8060944 0.7654986 0.7201265
#> [23,] 0.5478726 0.6758114 0.77601965 0.7985786 0.8130277 0.8916417 0.9039595
#> [24,] 0.6932233 0.5465987 0.46821011 0.3445773 0.5724645 0.8552360 0.9099883
#> [25,] 0.5043017 0.6815016 0.80102991 0.8324102 0.7226664 0.6342417 0.6687527
#>             V28       V29          V3       V30       V31       V32       V33
#>  [1,] 0.8271038 0.8111591 0.063781927 0.8319689 0.8162647 0.7620422 0.5797312
#>  [2,] 0.9663580 0.8532836 0.032349030 0.7136333 0.7487005 0.6698303 0.4734949
#>  [3,] 0.5997735 0.5010135 0.087652510 0.7374239 0.2936389 0.3534657 0.6482604
#>  [4,] 0.9694135 0.7432440 0.031152900 0.5639062 0.3564626 0.2046685 0.1874649
#>  [5,] 0.3307066 0.2870224 0.074403356 0.3097632 0.2495001 0.2598194 0.3307424
#>  [6,] 0.8012261 0.6703411 0.036195305 0.5130361 0.4125008 0.3650778 0.3427894
#>  [7,] 0.3992482 0.3829431 0.009865077 0.5487527 0.7098660 0.8242790 0.9247784
#>  [8,] 0.7936472 0.9476785 0.046392015 1.0035491 0.8213334 0.5698547 0.5384527
#>  [9,] 0.9578726 0.8435928 0.045284799 0.6328632 0.4277683 0.3137841 0.2796499
#> [10,] 0.8227161 0.7201836 0.081254346 0.5790308 0.5026480 0.3888674 0.2863229
#> [11,] 0.9867285 0.8846997 0.029687873 0.6330314 0.3850810 0.2836783 0.1991212
#> [12,] 0.6141670 0.6015107 0.034505959 0.4750106 0.3115514 0.2707768 0.2670729
#> [13,] 0.2848246 0.4369233 0.080712069 0.6181766 0.2715832 0.3305689 0.5223671
#> [14,] 1.0178306 0.9295013 0.030532319 0.6916032 0.6124591 0.4950118 0.3960427
#> [15,] 0.5530182 0.5536978 0.052114799 0.4081942 0.4618876 0.6089963 0.6471029
#> [16,] 0.8824636 0.7745198 0.020732071 0.6395653 0.5305145 0.3959635 0.2998835
#> [17,] 0.6895676 0.3652468 0.022718397 0.3220946 0.3743489 0.3192717 0.2843798
#> [18,] 0.6763596 0.7596741 0.033064033 0.7612800 0.5711828 0.5377132 0.7451679
#> [19,] 0.8331939 0.9776958 0.013819549 0.9523798 0.8878370 0.8433875 0.8781935
#> [20,] 0.4580465 0.4790571 0.016827593 0.5746956 0.4256745 0.2626090 0.3103512
#> [21,] 0.5452289 0.7114407 0.061948544 0.7497581 0.7729898 0.7076823 0.6467964
#> [22,] 0.7052818 0.5857368 0.014276503 0.4051991 0.2627791 0.2227353 0.2182679
#> [23,] 0.8584246 0.8679915 0.014996664 0.7939840 0.7057923 0.6306254 0.5791440
#> [24,] 0.5228380 0.3260436 0.056757672 0.1841219 0.2005541 0.4128191 0.4590197
#> [25,] 0.5991144 0.4042726 0.072251278 0.4265990 0.4813856 0.3645594 0.2653489
#>             V34        V35        V36        V37        V38        V39
#>  [1,] 0.4567586 0.39125572 0.31919899 0.33497355 0.39343482 0.41907653
#>  [2,] 0.1913059 0.09782064 0.11012710 0.13053887 0.14491387 0.17374835
#>  [3,] 0.8371032 0.89690967 0.92089520 0.86188001 0.93885108 0.70066407
#>  [4,] 0.4220168 0.68633570 0.79900025 0.79090147 0.67483726 0.59734155
#>  [5,] 0.3130575 0.33503997 0.26275865 0.29934992 0.35669855 0.31804687
#>  [6,] 0.2591959 0.18601851 0.22464031 0.27809928 0.30872762 0.37587138
#>  [7,] 0.8644391 0.65901837 0.43411188 0.11148374 0.15359339 0.36354234
#>  [8,] 0.4953715 0.31492540 0.31215592 0.26333091 0.11028677 0.19256909
#>  [9,] 0.2228294 0.14596997 0.04362251 0.08180591 0.14696372 0.12376610
#> [10,] 0.2416233 0.18016117 0.10265893 0.20523519 0.23465913 0.27002093
#> [11,] 0.1688805 0.17520662 0.10427792 0.06946313 0.08804392 0.08025028
#> [12,] 0.2180905 0.20001246 0.17560485 0.11670238 0.13559321 0.19061580
#> [13,] 0.6500049 0.81487547 0.82558331 0.77257695 0.87119145 0.83204392
#> [14,] 0.3235661 0.13307794 0.04472938 0.17325471 0.19117618 0.26175197
#> [15,] 0.6702348 0.63100018 0.61698950 0.55914081 0.31018494 0.21505745
#> [16,] 0.1877715 0.19385020 0.30545688 0.26117579 0.25401441 0.26517064
#> [17,] 0.2283327 0.16147927 0.06360725 0.15814826 0.15096772 0.10279384
#> [18,] 0.8930471 0.84807278 0.68629760 0.42820933 0.29322682 0.24120931
#> [19,] 0.8187093 0.67610637 0.50579498 0.34243511 0.23469830 0.14711260
#> [20,] 0.3259765 0.15732117 0.09145855 0.03987722 0.06721753 0.18389283
#> [21,] 0.6934631 0.75880332 0.82482325 0.82343653 0.85567319 0.77286171
#> [22,] 0.2694992 0.20720329 0.25637606 0.28119849 0.22734855 0.27109799
#> [23,] 0.6152354 0.65930696 0.65288567 0.58225037 0.46963830 0.32768727
#> [24,] 0.3477812 0.33468734 0.42207170 0.56602047 0.40594944 0.51145694
#> [25,] 0.1880369 0.43825622 0.63561491 0.65013974 0.42374648 0.26044107
#>                V4        V40        V41        V42        V43        V44
#>  [1,] 0.056256437 0.37162965 0.53161093 0.57312365 0.48813519 0.46013232
#>  [2,] 0.055915587 0.05140248 0.13855616 0.08352977 0.10979073 0.15899009
#>  [3,] 0.111036325 0.45877686 0.41634097 0.34212676 0.19805244 0.22204787
#>  [4,] 0.034386033 0.43314365 0.34281519 0.28788451 0.36776699 0.34882486
#>  [5,] 0.089295829 0.22915848 0.17466355 0.22911647 0.23633658 0.15008322
#>  [6,] 0.049330997 0.36751743 0.26213345 0.27253462 0.28734275 0.26301477
#>  [7,] 0.034776547 0.54887927 0.54310650 0.40458097 0.24167997 0.09751813
#>  [8,] 0.060566624 0.28630840 0.37000032 0.40725927 0.34329729 0.09284316
#>  [9,] 0.031413944 0.13657480 0.04463112 0.13475183 0.23678666 0.25512485
#> [10,] 0.129613948 0.31556788 0.32346157 0.31435230 0.19427333 0.14548547
#> [11,] 0.039590321 0.07051203 0.06603066 0.10596138 0.18657416 0.21742774
#> [12,] 0.053688350 0.20320783 0.20432855 0.19306468 0.18921574 0.21837710
#> [13,] 0.042010897 0.58805435 0.27980751 0.10647505 0.13861830 0.25789742
#> [14,] 0.049667799 0.27984427 0.26589831 0.27098814 0.10209662 0.07657080
#> [15,] 0.050855359 0.43522320 0.48855326 0.51447654 0.39069875 0.30082774
#> [16,] 0.028376457 0.18379517 0.15911606 0.19768965 0.16791941 0.13219981
#> [17,] 0.034644598 0.13040722 0.09544136 0.04740934 0.06095377 0.12287404
#> [18,] 0.040111236 0.39750948 0.39707361 0.31260237 0.32787130 0.21643923
#> [19,] 0.012470522 0.23397577 0.25420416 0.28301128 0.27031949 0.27532208
#> [20,] 0.058026395 0.21304748 0.19823481 0.09976575 0.04763745 0.09371969
#> [21,] 0.033236887 0.64852867 0.58838070 0.51277487 0.42969549 0.35773717
#> [22,] 0.006276469 0.23659623 0.18802718 0.17338399 0.15566996 0.16331414
#> [23,] 0.019298082 0.25763875 0.22521063 0.12773530 0.13831256 0.11622737
#> [24,] 0.033678009 0.46946967 0.28662428 0.25863013 0.25322745 0.23192194
#> [25,] 0.064278118 0.35480848 0.38088474 0.34175464 0.23974391 0.18619873
#>              V45        V46        V47        V48        V49         V5
#>  [1,] 0.54935702 0.48267850 0.31707145 0.21113220 0.11759641 0.09844093
#>  [2,] 0.10086108 0.04227625 0.09027846 0.11226639 0.06094411 0.06271535
#>  [3,] 0.34983648 0.24719936 0.15689783 0.17996213 0.08078245 0.06076618
#>  [4,] 0.42396461 0.36944892 0.23021168 0.11711224 0.08554443 0.02899789
#>  [5,] 0.09815381 0.07360973 0.06783307 0.05005837 0.03601096 0.10809065
#>  [6,] 0.23986432 0.14481497 0.11027283 0.09984477 0.06098721 0.06822095
#>  [7,] 0.08184640 0.11219476 0.12722578 0.11776105 0.08341915 0.04258662
#>  [8,] 0.13062114 0.06031935 0.08428631 0.10414041 0.06351355 0.07754487
#>  [9,] 0.22360583 0.21339661 0.16796475 0.10336326 0.05632799 0.06965535
#> [10,] 0.19256419 0.20175967 0.17565586 0.13138561 0.05410819 0.13318467
#> [11,] 0.21326901 0.22465027 0.17930350 0.11272031 0.05648237 0.08606293
#> [12,] 0.13836565 0.09163765 0.09054629 0.07506498 0.06050021 0.09229883
#> [13,] 0.49673032 0.39425722 0.20888313 0.17841556 0.10833047 0.03276478
#> [14,] 0.12712246 0.17519679 0.19249401 0.15336960 0.05494401 0.04072611
#> [15,] 0.19209801 0.14745237 0.08892974 0.06708597 0.02879983 0.06257246
#> [16,] 0.10980717 0.09407048 0.09234154 0.06671534 0.02745087 0.04487567
#> [17,] 0.09499455 0.05135936 0.07082474 0.05156740 0.03218157 0.05360832
#> [18,] 0.13382740 0.11065384 0.08658684 0.05201405 0.02097988 0.06883122
#> [19,] 0.25535593 0.17746476 0.08911163 0.04974730 0.03270864 0.01664883
#> [20,] 0.08370650 0.02532135 0.05057386 0.04051415 0.02532350 0.08357771
#> [21,] 0.20900760 0.15238092 0.13624782 0.11440738 0.05706907 0.05220112
#> [22,] 0.12288873 0.08696625 0.03528705 0.03503092 0.02402465 0.02026459
#> [23,] 0.10688673 0.07751135 0.04893157 0.03355266 0.01615235 0.03776176
#> [24,] 0.10754614 0.07824357 0.07632412 0.04976983 0.05770319 0.04252589
#> [25,] 0.18750155 0.15195998 0.14090895 0.12057466 0.08314883 0.06494052
#>               V50         V51         V52         V53         V54          V55
#>  [1,] 0.045888916 0.034564440 0.024578555 0.015239355 0.014075592 0.0128587792
#>  [2,] 0.020567162 0.007841955 0.008642469 0.006741287 0.010646292 0.0054919321
#>  [3,] 0.019700029 0.035443114 0.024848004 0.016277426 0.028571824 0.0184744413
#>  [4,] 0.028592254 0.012823997 0.014786283 0.005980163 0.002676196 0.0078835318
#>  [5,] 0.019931153 0.015463466 0.014713603 0.013428764 0.016283070 0.0116279940
#>  [6,] 0.019142451 0.018111018 0.012198139 0.011197867 0.012312590 0.0092728437
#>  [7,] 0.025898165 0.014005251 0.007557264 0.004233292 0.002034650 0.0075063136
#>  [8,] 0.004787488 0.011988876 0.015414686 0.005046100 0.011441090 0.0008477751
#>  [9,] 0.019361198 0.018191315 0.012692710 0.003960574 0.003638577 0.0045108944
#> [10,] 0.017012345 0.015505781 0.019335122 0.011771689 0.010388536 0.0148347335
#> [11,] 0.014886157 0.023548552 0.015661601 0.005759563 0.001943405 0.0033211598
#> [12,] 0.012783416 0.017931686 0.016365310 0.009288531 0.009086586 0.0062551538
#> [13,] 0.050501251 0.022253441 0.009254067 0.006897497 0.013915902 0.0171962461
#> [14,] 0.013520378 0.016389291 0.014446380 0.006225867 0.007220092 0.0088639197
#> [15,] 0.010940789 0.004780418 0.004099321 0.010722133 0.008364664 0.0092796902
#> [16,] 0.016642569 0.007188381 0.010200362 0.010085760 0.009533239 0.0070425040
#> [17,] 0.019508585 0.012324409 0.011703512 0.009664687 0.004161098 0.0027994961
#> [18,] 0.022662579 0.021395687 0.017622404 0.017330907 0.014055148 0.0067851372
#> [19,] 0.012366270 0.008158326 0.008313020 0.006299188 0.007259083 0.0036010552
#> [20,] 0.009243884 0.014571221 0.003552279 0.006515826 0.010788488 0.0092172674
#> [21,] 0.018305569 0.012391371 0.011881639 0.019557248 0.007542329 0.0073854879
#> [22,] 0.004680638 0.008467650 0.017118003 0.008951345 0.004826721 0.0028408348
#> [23,] 0.015673584 0.004941483 0.005695629 0.006862783 0.008275974 0.0063568927
#> [24,] 0.022986471 0.006787360 0.008718766 0.005010197 0.019052674 0.0194844576
#> [25,] 0.026356045 0.015922604 0.014061425 0.008070502 0.011593282 0.0070572386
#>               V56          V57         V58         V59         V6         V60
#>  [1,] 0.012593833 0.0110674335 0.012247269 0.009075602 0.10840095 0.006335346
#>  [2,] 0.006581842 0.0053611838 0.005925799 0.006049107 0.10619151 0.006378516
#>  [3,] 0.022766917 0.0157352177 0.016066252 0.019323435 0.12656083 0.015538654
#>  [4,] 0.003557794 0.0036744912 0.004355153 0.005322416 0.06433420 0.005716036
#>  [5,] 0.012625431 0.0097028018 0.013218626 0.012435758 0.09572776 0.009721659
#>  [6,] 0.009485174 0.0068537655 0.009088739 0.007648444 0.10865685 0.005327931
#>  [7,] 0.005241202 0.0027004064 0.007170738 0.007317467 0.07388408 0.003808470
#>  [8,] 0.003844351 0.0098945435 0.006404438 0.004352540 0.12955179 0.008455543
#>  [9,] 0.001774764 0.0031545902 0.001582606 0.002674704 0.04642738 0.003310444
#> [10,] 0.003880424 0.0098031506 0.009048499 0.006162444 0.09944607 0.010061979
#> [11,] 0.006167884 0.0029582640 0.004787667 0.001513541 0.09573080 0.003895075
#> [12,] 0.004810407 0.0059542283 0.012945145 0.008817658 0.16107365 0.003195762
#> [13,] 0.013534046 0.0174100103 0.015590615 0.010327790 0.12520938 0.007482038
#> [14,] 0.003854586 0.0044836693 0.004957732 0.003579543 0.05718408 0.002382463
#> [15,] 0.012822113 0.0119585276 0.003021849 0.009838110 0.12340063 0.009946521
#> [16,] 0.007892653 0.0045541246 0.003900774 0.004220015 0.06123186 0.003437627
#> [17,] 0.009184528 0.0069296948 0.002349986 0.004899648 0.08928723 0.003824718
#> [18,] 0.010883847 0.0176460972 0.006684375 0.007536938 0.08108260 0.010942042
#> [19,] 0.002066805 0.0019705249 0.002552825 0.006228856 0.03219542 0.003804195
#> [20,] 0.005718465 0.0004630698 0.002218465 0.002538849 0.10596501 0.003413430
#> [21,] 0.007300749 0.0042041058 0.014041927 0.012148020 0.08051908 0.009091299
#> [22,] 0.002909397 0.0037707619 0.008659981 0.004703059 0.05302230 0.003989903
#> [23,] 0.005121875 0.0049306670 0.003298676 0.003363281 0.07642976 0.003569621
#> [24,] 0.005407328 0.0122955462 0.020602963 0.014642058 0.01561187 0.004788858
#> [25,] 0.005172520 0.0105305585 0.006675977 0.007802666 0.10298016 0.006209511
#>               V7         V8         V9
#>  [1,] 0.11894829 0.14422572 0.23141712
#>  [2,] 0.11371823 0.12148720 0.14584433
#>  [3,] 0.26422874 0.45328016 0.63381107
#>  [4,] 0.04638812 0.05026251 0.03054894
#>  [5,] 0.10675237 0.08913988 0.14636596
#>  [6,] 0.12367489 0.16282688 0.25090877
#>  [7,] 0.07574825 0.16281509 0.20817299
#>  [8,] 0.06076373 0.05294271 0.17516439
#>  [9,] 0.08741761 0.10705712 0.15749548
#> [10,] 0.11296211 0.09432763 0.11574951
#> [11,] 0.11209169 0.14043045 0.20805257
#> [12,] 0.17968990 0.18832095 0.23255283
#> [13,] 0.20510539 0.30490569 0.46933134
#> [14,] 0.11756745 0.06859188 0.11425244
#> [15,] 0.12216737 0.13536255 0.05516205
#> [16,] 0.08837303 0.08365145 0.08940946
#> [17,] 0.13677714 0.10400409 0.10658719
#> [18,] 0.11402111 0.10144953 0.06414904
#> [19,] 0.06263994 0.02948930 0.04897734
#> [20,] 0.11503479 0.06213983 0.05570001
#> [21,] 0.08011747 0.11017652 0.16322582
#> [22,] 0.07878800 0.12182375 0.09276092
#> [23,] 0.09378699 0.09262026 0.12574018
#> [24,] 0.02560969 0.00584026 0.06752809
#> [25,] 0.13772591 0.19871489 0.24693007
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841