Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.010272967 0.11882410 0.19584078 0.22134019 0.19817172 0.1412189
#>  [2,] 0.027720432 0.24923338 0.21145790 0.27429985 0.29285905 0.2301475
#>  [3,] 0.011890166 0.33800022 0.33035738 0.33209071 0.24537304 0.1881267
#>  [4,] 0.058465673 0.16148643 0.30454273 0.44532547 0.52815551 0.5417726
#>  [5,] 0.042437590 0.53797950 0.37828518 0.10208228 0.25139215 0.3469060
#>  [6,] 0.025908405 0.18168763 0.22146783 0.25354140 0.30158829 0.3930825
#>  [7,] 0.027237496 0.56762707 0.64578558 0.55510074 0.51325246 0.3316946
#>  [8,] 0.032587490 0.25582222 0.31662817 0.31934322 0.27286748 0.1343931
#>  [9,] 0.027257178 0.14516314 0.23650410 0.26203919 0.26364746 0.2434336
#> [10,] 0.029976152 0.39426075 0.48683474 0.47037792 0.38823134 0.2957739
#> [11,] 0.020590221 0.29554091 0.34307091 0.42502831 0.43450156 0.3765686
#> [12,] 0.016579463 0.24406617 0.28146799 0.32802752 0.30979137 0.2499161
#> [13,] 0.106145566 0.60829164 0.48648757 0.23119168 0.09956643 0.3175339
#> [14,] 0.016729409 0.09741965 0.10933270 0.10878187 0.12829642 0.1633242
#> [15,] 0.017274455 0.09176166 0.06880231 0.11448646 0.27568873 0.2924003
#> [16,] 0.024690554 0.09222028 0.12830642 0.18671339 0.17271279 0.1494487
#> [17,] 0.021200165 0.21499327 0.23331146 0.12866480 0.19230556 0.4731479
#> [18,] 0.008798522 0.08352949 0.08978118 0.08831762 0.13606461 0.1434446
#> [19,] 0.016494396 0.10028300 0.12475433 0.14729463 0.19415773 0.2173006
#> [20,] 0.016400000 0.02510000 0.08010000 0.10560000 0.12660000 0.0890000
#> [21,] 0.085600000 0.27510000 0.27070000 0.09460000 0.10200000 0.4519000
#> [22,] 0.016829219 0.02555126 0.05455294 0.13498389 0.21784589 0.3544978
#> [23,] 0.029369832 0.28275734 0.18766536 0.16088025 0.10703504 0.3229230
#> [24,] 0.015168345 0.15819804 0.13588443 0.12271691 0.19522526 0.3647626
#> [25,] 0.018450657 0.10956286 0.10422995 0.11211700 0.12300485 0.1479283
#>              V15       V16       V17       V18        V19          V2       V20
#>  [1,] 0.05646359 0.1282352 0.2636185 0.3873466 0.49350270 0.014921344 0.5871145
#>  [2,] 0.14075841 0.1875549 0.2080798 0.1965109 0.19482575 0.026702285 0.2135664
#>  [3,] 0.10971203 0.2075358 0.2698564 0.3340770 0.23764689 0.037337460 0.1223234
#>  [4,] 0.62727019 0.7244079 0.8622604 0.9378125 0.98491883 0.073833055 0.9062248
#>  [5,] 0.42323737 0.4219577 0.1659928 0.1481722 0.32126278 0.079189541 0.9214677
#>  [6,] 0.46207851 0.5513684 0.6363271 0.6976348 0.76952895 0.029544804 0.8487681
#>  [7,] 0.17808725 0.2382793 0.2589343 0.3126233 0.57096083 0.033862845 0.7391360
#>  [8,] 0.13070174 0.1694061 0.2012336 0.2510580 0.21343017 0.022312776 0.2082215
#>  [9,] 0.21521645 0.1997404 0.2177276 0.2628438 0.37522818 0.061667589 0.3447343
#> [10,] 0.24651207 0.3338011 0.3592941 0.3950350 0.45973472 0.040815812 0.5761834
#> [11,] 0.33193745 0.3485706 0.3088875 0.2383852 0.37602740 0.044898851 0.5932881
#> [12,] 0.20028404 0.2391989 0.2594700 0.3766056 0.57626679 0.027229846 0.6824421
#> [13,] 0.47717951 0.4676186 0.1509721 0.1059564 0.34466533 0.122188210 0.9206756
#> [14,] 0.21657162 0.1906657 0.1710718 0.1688773 0.08446804 0.017020047 0.1301572
#> [15,] 0.22348091 0.2372834 0.2047748 0.3518242 0.67528954 0.037846104 0.8678837
#> [16,] 0.16946398 0.2481764 0.3518170 0.3641975 0.31824988 0.039326659 0.3512672
#> [17,] 0.67470423 0.7534129 0.8038673 0.6769352 0.48343311 0.033261272 0.3131590
#> [18,] 0.12698797 0.1353564 0.1441300 0.1914947 0.20811340 0.007910722 0.2155320
#> [19,] 0.22431491 0.2521337 0.3217782 0.3653677 0.37968337 0.014525923 0.4536552
#> [20,] 0.01980000 0.1133000 0.2826000 0.3234000 0.32380000 0.017300000 0.4333000
#> [21,] 0.67370000 0.6699000 0.7066000 0.5632000 0.37850000 0.045400000 0.2721000
#> [22,] 0.45610469 0.5878325 0.6872921 0.7540319 0.76500379 0.025930811 0.8132861
#> [23,] 0.42267298 0.4927255 0.4542532 0.4885733 0.41676496 0.068426926 0.5293375
#> [24,] 0.58796769 0.7880451 0.8933745 0.8519044 0.83780295 0.007500550 0.8132965
#> [25,] 0.16482921 0.1876599 0.1557991 0.1602149 0.13473845 0.017548546 0.1402999
#>              V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.67290784 0.7631907 0.7625161 0.7071115 0.7744154 0.9253759 1.0021887
#>  [2,] 0.17825397 0.3233199 0.4917412 0.6039158 0.6532406 0.6687339 0.7776890
#>  [3,] 0.05388511 0.1311528 0.3560289 0.5047919 0.6518435 0.8495059 0.9829972
#>  [4,] 0.82496990 0.7321731 0.6606932 0.5445040 0.3538242 0.2039451 0.1365569
#>  [5,] 0.97641347 0.3627071 0.1694827 0.1906979 0.1141754 0.3309535 0.2586274
#>  [6,] 0.89661598 0.9249991 0.9398343 0.9306784 0.8689177 0.8329857 0.7646162
#>  [7,] 0.88831754 0.8340759 0.7391231 0.7683687 0.7350763 0.8420593 0.9084284
#>  [8,] 0.32963663 0.6095399 0.8138786 0.8774511 0.9474375 0.9626100 0.9511091
#>  [9,] 0.41021633 0.4646841 0.5526833 0.6925682 0.7335557 0.7514249 0.8010368
#> [10,] 0.74971223 0.9268028 0.9519125 0.9339158 0.9054510 0.8157855 0.7409883
#> [11,] 0.69204020 0.6433623 0.5737832 0.6417813 0.6721021 0.7958268 0.9118764
#> [12,] 0.79101174 0.7878554 0.8102760 0.9212708 0.9355626 0.9562823 0.9757638
#> [13,] 0.97020595 0.4808850 0.1258072 0.1386606 0.2683195 0.5490723 0.3466587
#> [14,] 0.20601903 0.1697143 0.2815425 0.5514541 0.8621962 1.0208461 0.8739244
#> [15,] 0.83140695 0.7109549 0.5159321 0.4425042 0.5275640 0.7215856 0.8886123
#> [16,] 0.34466261 0.3386794 0.3934501 0.4825315 0.4881449 0.5091360 0.5596107
#> [17,] 0.49323108 0.7316278 0.8390828 0.9707858 0.8207446 0.6784418 0.6893437
#> [18,] 0.27471376 0.4344605 0.5418422 0.6553476 0.7118189 0.7805175 0.7814766
#> [19,] 0.53926985 0.6108225 0.6905715 0.7906243 0.8905705 0.9052835 0.9119343
#> [20,] 0.60680000 0.7652000 0.9203000 0.9719000 0.9207000 0.7545000 0.8289000
#> [21,] 0.52970000 0.7697000 0.8643000 0.9304000 0.9372000 0.6247000 0.6024000
#> [22,] 0.93484562 0.8790413 0.8421629 0.8941773 0.9031105 0.9072261 0.8053390
#> [23,] 0.69843833 0.8129287 0.8133606 0.5807842 0.5293395 0.5901295 0.5386632
#> [24,] 0.85544944 0.8931568 0.8539686 0.8908457 0.9452864 0.9948827 0.9209874
#> [25,] 0.23825060 0.2827993 0.4281457 0.6397820 0.8914125 0.9937898 0.8118708
#>             V28       V29          V3       V30       V31       V32       V33
#>  [1,] 0.9921482 0.8665638 0.030077738 0.6265024 0.2611059 0.1845021 0.2673258
#>  [2,] 0.8516125 0.8571946 0.052846399 0.9072537 0.9364004 0.8927964 0.7328893
#>  [3,] 0.9855027 0.9769445 0.030643207 1.0419038 0.9558946 0.8300211 0.7195421
#>  [4,] 0.1211767 0.1758422 0.072307394 0.2427866 0.2387275 0.2939821 0.3254463
#>  [5,] 0.5337466 0.4701042 0.100398110 0.7178147 0.2244475 0.2979058 0.6230797
#>  [6,] 0.6559277 0.5375061 0.042923633 0.4518318 0.3469275 0.3014546 0.2321335
#>  [7,] 0.9792486 0.8236086 0.045809018 0.4742152 0.3098997 0.2947850 0.2840546
#>  [8,] 1.0000000 0.8714721 0.029331842 0.6880373 0.7293639 0.6845956 0.5087155
#>  [9,] 0.8223951 0.7721200 0.081472561 0.6641550 0.5960900 0.4830341 0.4075042
#> [10,] 0.6232384 0.4205678 0.034868342 0.3383428 0.3288629 0.3880828 0.4581411
#> [11,] 0.9270219 0.8577519 0.058324229 0.7165825 0.6592424 0.6693814 0.4903646
#> [12,] 0.9711642 0.8688179 0.032850261 0.6503259 0.4302884 0.2987448 0.2439156
#> [13,] 0.6069865 0.5621781 0.083967826 0.8497160 0.5320134 0.5053175 0.5113225
#> [14,] 0.6767000 0.5579363 0.021359401 0.4456363 0.3496997 0.1988363 0.1594938
#> [15,] 0.7572993 0.5150376 0.038729664 0.3464260 0.3843566 0.3456057 0.3175604
#> [16,] 0.6955669 0.7895293 0.030724532 0.7582420 0.7454322 0.6772631 0.6594563
#> [17,] 0.5591649 0.4574129 0.085619343 0.5146787 0.4081884 0.1711078 0.1743230
#> [18,] 0.8521322 0.9279265 0.018596878 0.8810107 0.7630311 0.7028978 0.7249398
#> [19,] 0.8444046 0.7179757 0.021221818 0.6098134 0.5511484 0.4664679 0.3690068
#> [20,] 0.8907000 0.7309000 0.034700000 0.6896000 0.5829000 0.4935000 0.3101000
#> [21,] 0.6810000 0.5047000 0.038200000 0.5775000 0.4754000 0.2400000 0.2779000
#> [22,] 0.5878995 0.3460705 0.008416062 0.3321264 0.3175833 0.2072225 0.2166307
#> [23,] 0.4743267 0.2477239 0.089716958 0.2604703 0.5401695 0.6239415 0.3973849
#> [24,] 0.7261202 0.5683787 0.024807783 0.4833504 0.4723847 0.4295204 0.3641150
#> [25,] 0.6069220 0.5162576 0.012064975 0.3654328 0.2742878 0.1084073 0.1425088
#>              V34        V35        V36        V37        V38        V39
#>  [1,] 0.45977043 0.71543596 0.84866240 0.80351616 0.66552811 0.49853314
#>  [2,] 0.60764641 0.59838903 0.54633626 0.50488095 0.50420886 0.51373733
#>  [3,] 0.65731621 0.68880472 0.73262656 0.71759867 0.56335195 0.44013342
#>  [4,] 0.43769211 0.47298850 0.28941122 0.33375763 0.39062504 0.26290952
#>  [5,] 0.82786883 0.87529281 0.91929593 0.83636824 0.94429833 0.72050357
#>  [6,] 0.18268634 0.22937334 0.27779863 0.28225486 0.27286937 0.27277560
#>  [7,] 0.31027165 0.16197113 0.29801958 0.17807778 0.30741976 0.58220278
#>  [8,] 0.20028102 0.09514356 0.10966381 0.12049703 0.13957152 0.17961067
#>  [9,] 0.33970631 0.23100176 0.18576891 0.23543103 0.19972376 0.24730717
#> [10,] 0.44997556 0.29454933 0.32786872 0.35733815 0.31106477 0.37447452
#> [11,] 0.29250469 0.09771052 0.26547629 0.28173924 0.36350601 0.51328603
#> [12,] 0.18575573 0.15390252 0.07456306 0.06952288 0.08786939 0.11046252
#> [13,] 0.65016796 0.84637097 0.82234478 0.81816539 0.89389743 0.76454166
#> [14,] 0.20626326 0.31447693 0.23418109 0.09872961 0.04220734 0.06065205
#> [15,] 0.27753107 0.29598744 0.33618412 0.45236461 0.42531411 0.39231024
#> [16,] 0.75701220 0.80294099 0.80252510 0.73534143 0.60251585 0.48973759
#> [17,] 0.07150651 0.40407634 0.68227321 0.67092032 0.30544869 0.09127225
#> [18,] 0.65770482 0.58080055 0.49497582 0.34856428 0.29406383 0.27862067
#> [19,] 0.30025642 0.30863424 0.35231578 0.30683722 0.28673814 0.22995634
#> [20,] 0.03060000 0.02440000 0.11080000 0.15940000 0.13710000 0.06960000
#> [21,] 0.19970000 0.53050000 0.74090000 0.77750000 0.44240000 0.14160000
#> [22,] 0.20113953 0.13279368 0.05059238 0.07179572 0.11189508 0.15946842
#> [23,] 0.29029066 0.37845515 0.51176843 0.61146342 0.52566586 0.57819315
#> [24,] 0.32787532 0.24020885 0.15560672 0.26600405 0.20232893 0.12485611
#> [25,] 0.16795881 0.20190302 0.17768140 0.04552864 0.10179624 0.17759646
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.04330471 0.35804859 0.32956161 0.27729960 0.35256729 0.38592274
#>  [2,] 0.04930009 0.52425703 0.58697151 0.48305002 0.43186879 0.54945955
#>  [3,] 0.07519916 0.48280999 0.43744194 0.38784994 0.47281968 0.55644211
#>  [4,] 0.07570773 0.23236634 0.30432549 0.34053143 0.23279048 0.10709424
#>  [5,] 0.10151044 0.47416670 0.43202304 0.30086328 0.14580540 0.18620937
#>  [6,] 0.05552091 0.28671189 0.29312471 0.29943446 0.22058425 0.19393067
#>  [7,] 0.06053955 0.42361661 0.10897332 0.34461614 0.45048385 0.47835008
#>  [8,] 0.05183026 0.05700571 0.15261763 0.07443322 0.10913490 0.15466003
#>  [9,] 0.09959550 0.32168987 0.33975727 0.33913440 0.25722548 0.15660609
#> [10,] 0.04652326 0.37422233 0.28571472 0.18825771 0.31458892 0.30209286
#> [11,] 0.04219346 0.43386202 0.40330803 0.54070606 0.52864331 0.50168480
#> [12,] 0.03465805 0.12592341 0.07278608 0.11195127 0.20107495 0.22784085
#> [13,] 0.13424622 0.49586815 0.37067652 0.44412907 0.45674621 0.44108706
#> [14,] 0.01841491 0.13916982 0.23207561 0.20383549 0.16609360 0.09089447
#> [15,] 0.03529351 0.29460413 0.17748370 0.15967129 0.11726224 0.08329461
#> [16,] 0.02637961 0.40056989 0.31680553 0.27614835 0.22554877 0.20375623
#> [17,] 0.07207596 0.32040513 0.42958472 0.37441807 0.22980113 0.18454940
#> [18,] 0.02483092 0.26361834 0.20698061 0.15196292 0.18269431 0.16965975
#> [19,] 0.02392386 0.19249002 0.17210988 0.18905858 0.17039577 0.12917631
#> [20,] 0.00700000 0.04520000 0.06200000 0.14210000 0.15970000 0.13840000
#> [21,] 0.02030000 0.35080000 0.44820000 0.42080000 0.30540000 0.22350000
#> [22,] 0.04592914 0.16563056 0.14890664 0.06422024 0.07278771 0.12204334
#> [23,] 0.10264415 0.60325735 0.48677559 0.31031092 0.25985390 0.23459992
#> [24,] 0.02531541 0.14769891 0.08275957 0.08270996 0.10049977 0.08997498
#> [25,] 0.01940782 0.24762356 0.31227279 0.29398791 0.19632801 0.08248032
#>              V45        V46        V47        V48        V49         V5
#>  [1,] 0.33658549 0.23497942 0.11731082 0.05177103 0.04135563 0.03522509
#>  [2,] 0.53790008 0.45289996 0.33703017 0.21473093 0.10172872 0.06325249
#>  [3,] 0.47640607 0.29041787 0.16511465 0.15411984 0.10061635 0.10678795
#>  [4,] 0.07316973 0.08898590 0.07223897 0.03957514 0.03882656 0.09579696
#>  [5,] 0.34118876 0.24655954 0.18336891 0.22364286 0.10456694 0.04180897
#>  [6,] 0.13699413 0.12055221 0.11602002 0.08869626 0.05472524 0.11167289
#>  [7,] 0.36922979 0.11353559 0.11393556 0.17243143 0.10651085 0.12527479
#>  [8,] 0.08926432 0.03838674 0.09006680 0.12011500 0.06714836 0.07100669
#>  [9,] 0.16492747 0.17816273 0.17384862 0.15162661 0.07333366 0.10559115
#> [10,] 0.29004261 0.21692563 0.14676293 0.10087271 0.06340272 0.03526232
#> [11,] 0.56557069 0.36025624 0.24785817 0.22636394 0.12693443 0.08044500
#> [12,] 0.21848715 0.21047443 0.16785943 0.10709801 0.05247985 0.07009892
#> [13,] 0.56687882 0.43800143 0.08997981 0.08805357 0.03385783 0.13237903
#> [14,] 0.08218459 0.07126115 0.03350902 0.04397629 0.03412521 0.03950555
#> [15,] 0.09194779 0.09442584 0.08696617 0.06203107 0.03787521 0.02651732
#> [16,] 0.15256351 0.10568344 0.08478433 0.05935602 0.03096495 0.04438432
#> [17,] 0.20984584 0.16904293 0.17789099 0.15959345 0.11752672 0.04020958
#> [18,] 0.12773690 0.07385872 0.04038205 0.03801907 0.01990651 0.02792789
#> [19,] 0.10251209 0.07422934 0.06849387 0.05169648 0.02398296 0.04453193
#> [20,] 0.03720000 0.06880000 0.08670000 0.05130000 0.00920000 0.01870000
#> [21,] 0.26110000 0.27980000 0.23920000 0.20210000 0.13260000 0.03850000
#> [22,] 0.08239592 0.04636647 0.06796456 0.04940651 0.03792816 0.05158955
#> [23,] 0.18039448 0.11893252 0.12918917 0.09247923 0.04458043 0.12197105
#> [24,] 0.12452210 0.11043257 0.08606474 0.06554347 0.03386348 0.06427666
#> [25,] 0.08145157 0.09540296 0.04684364 0.02586218 0.02668072 0.05454341
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.012388977 0.003585810 0.009438947 0.007781244 0.002214576 0.002753337
#>  [2,] 0.039609555 0.028666231 0.020247234 0.007173453 0.014123193 0.014107033
#>  [3,] 0.027772507 0.011481627 0.016788094 0.009801558 0.008971300 0.001392410
#>  [4,] 0.025802244 0.013453596 0.026108394 0.016062575 0.022327409 0.021715668
#>  [5,] 0.026035115 0.039937706 0.020697803 0.014943492 0.029395860 0.017551345
#>  [6,] 0.019965313 0.016084957 0.017380205 0.015271374 0.015855820 0.007187144
#>  [7,] 0.027564650 0.033903557 0.016748357 0.017454116 0.013526965 0.007636055
#>  [8,] 0.021610352 0.010012723 0.008685536 0.007718772 0.011728485 0.005589617
#>  [9,] 0.013003509 0.019625852 0.020479429 0.012549976 0.012424863 0.009129994
#> [10,] 0.023248013 0.015375016 0.007018840 0.009106963 0.014487478 0.011813618
#> [11,] 0.034400316 0.036442324 0.021492526 0.009148855 0.014493778 0.012883240
#> [12,] 0.016564009 0.019387490 0.013542559 0.008117741 0.005463738 0.003701202
#> [13,] 0.019969527 0.014858317 0.044536773 0.024196724 0.011724529 0.013625497
#> [14,] 0.011324575 0.008121465 0.013245953 0.004214774 0.004354581 0.005523583
#> [15,] 0.024051565 0.013651648 0.012641382 0.017029567 0.020527759 0.013475372
#> [16,] 0.014904809 0.010068003 0.009903193 0.013141620 0.008392944 0.011693977
#> [17,] 0.029861309 0.018120869 0.009473214 0.004712851 0.010468933 0.004683022
#> [18,] 0.009982387 0.008611520 0.008936810 0.011280598 0.005888665 0.005504100
#> [19,] 0.012826573 0.006056320 0.008058446 0.007072361 0.007419208 0.006933019
#> [20,] 0.019800000 0.011800000 0.009000000 0.022300000 0.017900000 0.008400000
#> [21,] 0.035800000 0.012800000 0.017200000 0.013800000 0.007900000 0.003700000
#> [22,] 0.023330436 0.014933023 0.007673349 0.007879116 0.005088110 0.007061177
#> [23,] 0.020179282 0.020013471 0.019790073 0.018511076 0.013179278 0.007660654
#> [24,] 0.010193059 0.016700345 0.010241045 0.005787269 0.008915598 0.005767732
#> [25,] 0.009582145 0.012077810 0.007317855 0.005675699 0.007212835 0.008124986
#>               V56          V57         V58          V59         V6         V60
#>  [1,] 0.003163141 0.0009095727 0.000765246 0.0027318164 0.04060893 0.003237339
#>  [2,] 0.010711192 0.0104112676 0.004184678 0.0109775515 0.07870244 0.006361180
#>  [3,] 0.007927422 0.0062272614 0.005170840 0.0012126192 0.11521807 0.002945167
#>  [4,] 0.010616268 0.0027919927 0.015022149 0.0149223944 0.05791473 0.006263705
#>  [5,] 0.025390707 0.0188694416 0.016121152 0.0195476624 0.12350520 0.014227134
#>  [6,] 0.010467639 0.0086419549 0.010013259 0.0076620098 0.12816264 0.005075150
#>  [7,] 0.003581382 0.0025182800 0.002851469 0.0058410326 0.20048359 0.005640251
#>  [8,] 0.007700203 0.0056201628 0.006081410 0.0081640529 0.10979644 0.007908226
#>  [9,] 0.008769833 0.0080771228 0.009170474 0.0081172388 0.12171991 0.005693680
#> [10,] 0.010209253 0.0072996471 0.011763889 0.0083933061 0.07395679 0.005805953
#> [11,] 0.004497806 0.0075821389 0.006661246 0.0052725386 0.10513441 0.003243647
#> [12,] 0.004429245 0.0044071350 0.003227365 0.0024541804 0.09741709 0.003933820
#> [13,] 0.018668052 0.0112776284 0.018628654 0.0157229639 0.15402003 0.017864970
#> [14,] 0.001391958 0.0032198968 0.002379313 0.0006727163 0.08946625 0.001394371
#> [15,] 0.009991384 0.0093903167 0.013413098 0.0043293375 0.01147123 0.003858318
#> [16,] 0.006476668 0.0072441954 0.007077776 0.0065309176 0.08952117 0.003193736
#> [17,] 0.005506967 0.0100224922 0.007759455 0.0102499825 0.06382455 0.006274621
#> [18,] 0.006291792 0.0056981621 0.002874007 0.0066119869 0.04233889 0.005311318
#> [19,] 0.006946743 0.0053821772 0.004832163 0.0045363558 0.06233120 0.004914741
#> [20,] 0.006800000 0.0032000000 0.003500000 0.0056000000 0.06710000 0.004000000
#> [21,] 0.005100000 0.0258000000 0.010200000 0.0037000000 0.05340000 0.003700000
#> [22,] 0.006534853 0.0035223325 0.002346835 0.0031201820 0.09534825 0.003238905
#> [23,] 0.010967741 0.0081860508 0.010912358 0.0041005260 0.13673996 0.008618424
#> [24,] 0.007028260 0.0062951540 0.003693746 0.0055132308 0.11763366 0.004945901
#> [25,] 0.004268568 0.0027671421 0.002179293 0.0041935541 0.05589487 0.002497148
#>               V7         V8         V9
#>  [1,] 0.04125982 0.01906727 0.07664509
#>  [2,] 0.09642702 0.11346380 0.21354763
#>  [3,] 0.15956886 0.21886024 0.34442690
#>  [4,] 0.07464656 0.03134844 0.04403709
#>  [5,] 0.26404203 0.43193709 0.62208901
#>  [6,] 0.13475811 0.12365318 0.16798988
#>  [7,] 0.20718432 0.23647720 0.34257412
#>  [8,] 0.12180025 0.13304448 0.15498851
#>  [9,] 0.10576646 0.07586661 0.12998390
#> [10,] 0.09054414 0.17383232 0.31420874
#> [11,] 0.11291298 0.17856489 0.25557375
#> [12,] 0.11344488 0.13838722 0.19489961
#> [13,] 0.18192554 0.41630893 0.68597102
#> [14,] 0.09877862 0.08769063 0.11731327
#> [15,] 0.03048830 0.03237585 0.08142123
#> [16,] 0.09688854 0.09953270 0.08610455
#> [17,] 0.13578563 0.19369823 0.19801576
#> [18,] 0.04392705 0.04456217 0.06570052
#> [19,] 0.07903831 0.07296825 0.09750470
#> [20,] 0.10560000 0.06970000 0.09620000
#> [21,] 0.21400000 0.31100000 0.28370000
#> [22,] 0.15737051 0.10941480 0.07728355
#> [23,] 0.09606306 0.14691698 0.18172980
#> [24,] 0.10841158 0.08308250 0.15868046
#> [25,] 0.05980354 0.05173360 0.09996788
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913