Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1         V10        V11        V12        V13        V14
#>  [1,] 0.01688159  0.32716791 0.39858586 0.46474294 0.48449099 0.41388324
#>  [2,] 0.03082162  0.24262698 0.29994323 0.29848720 0.24879845 0.14722619
#>  [3,] 0.04010720 -0.04275074 0.10913192 0.28152237 0.32233130 0.40545122
#>  [4,] 0.04813795  0.17239654 0.18531270 0.20159054 0.29408706 0.30921994
#>  [5,] 0.06634849  0.34589467 0.21493496 0.13664707 0.40978808 0.57426045
#>  [6,] 0.02887265  0.24661260 0.30299380 0.36052406 0.36676744 0.35493023
#>  [7,] 0.02185649  0.57562773 0.62108621 0.44936489 0.34462040 0.25594409
#>  [8,] 0.09431569  0.58328441 0.41106919 0.10974235 0.10979113 0.27440999
#>  [9,] 0.02793434  0.18043671 0.15587939 0.18939242 0.28987265 0.33582693
#> [10,] 0.09960966  0.37639759 0.47514050 0.44082017 0.64573331 0.85437296
#> [11,] 0.03955988  0.22931263 0.17318321 0.27665124 0.30883544 0.26976220
#> [12,] 0.02023223  0.27366296 0.30830056 0.34999204 0.29419651 0.20369952
#> [13,] 0.02357245  0.30226617 0.40902424 0.45793650 0.39410535 0.30906819
#> [14,] 0.02692089  0.23830443 0.27697386 0.24920751 0.16008569 0.02435929
#> [15,] 0.01899001  0.16127223 0.10783964 0.14452536 0.34265432 0.31781452
#> [16,] 0.01405726  0.13072625 0.13982023 0.12860843 0.10445977 0.14810593
#> [17,] 0.02889125  0.25278828 0.24824828 0.11598616 0.20031056 0.49433481
#> [18,] 0.01499742  0.03122416 0.03215322 0.01274845 0.09466557 0.12811133
#> [19,] 0.02380057  0.12974726 0.13384317 0.06568291 0.04350058 0.14468615
#> [20,] 0.03129092  0.11524933 0.28280511 0.38657760 0.31307938 0.24266538
#> [21,] 0.00730094  0.09212797 0.07331790 0.09415372 0.06272939 0.09309562
#> [22,] 0.00908368  0.11286631 0.10710721 0.07123975 0.09792060 0.09844608
#> [23,] 0.05843532  0.35266636 0.33772019 0.24721323 0.29847616 0.50034363
#> [24,] 0.02238855  0.09941607 0.13303454 0.05870779 0.11209385 0.14810176
#> [25,] 0.02136928  0.09035639 0.11950068 0.09817263 0.12313340 0.11213935
#>              V15        V16       V17       V18       V19           V2
#>  [1,] 0.32517712 0.33478939 0.3139584 0.2904020 0.3765638  0.032971997
#>  [2,] 0.10147915 0.13379090 0.2028102 0.2500585 0.2808060  0.038322157
#>  [3,] 0.55511835 0.66404107 0.8708674 0.8384237 0.8777131  0.058892843
#>  [4,] 0.29202319 0.29942610 0.3246368 0.1776705 0.2163656  0.066355074
#>  [5,] 0.24094538 0.43686657 0.6044072 0.5831769 0.8246469  0.065718374
#>  [6,] 0.36149563 0.43233982 0.5105568 0.6018068 0.6512308  0.049522040
#>  [7,] 0.17731524 0.14358462 0.2141189 0.3969540 0.6061714  0.029200172
#>  [8,] 0.41351380 0.39144742 0.2695080 0.2363378 0.4083052  0.102833867
#>  [9,] 0.38728715 0.46730988 0.5990973 0.7801564 0.9586685  0.057523834
#> [10,] 0.95171828 0.95730377 0.8402347 0.8750525 0.9128680  0.106042195
#> [11,] 0.32188659 0.46715020 0.4407505 0.1786473 0.4455788  0.103201413
#> [12,] 0.14119549 0.19975707 0.2416858 0.3861981 0.5908982  0.033366315
#> [13,] 0.26987857 0.36066041 0.4024656 0.3965990 0.4649795  0.024597288
#> [14,] 0.11040349 0.15683256 0.2107883 0.3067835 0.3570635  0.036223141
#> [15,] 0.18450100 0.16015660 0.0639632 0.3360657 0.8152921  0.030004817
#> [16,] 0.11028847 0.09965520 0.2014080 0.2961596 0.3393780 -0.001362772
#> [17,] 0.69037695 0.74508994 0.7845965 0.6287011 0.4254072  0.059204303
#> [18,] 0.18539716 0.24937816 0.3142236 0.3574996 0.4216263  0.017621365
#> [19,] 0.06959573 0.08635346 0.2324611 0.2492137 0.2865802  0.002854420
#> [20,] 0.28590820 0.44001804 0.6275995 0.5756248 0.5357769  0.061609504
#> [21,] 0.09262035 0.14273461 0.1681466 0.3002729 0.3535296  0.009811751
#> [22,] 0.15865869 0.19258491 0.2060633 0.2163122 0.2288042  0.010934328
#> [23,] 0.65585597 0.71173140 0.7343696 0.6334468 0.4507193  0.054423845
#> [24,] 0.18458523 0.17159596 0.1336849 0.1458180 0.1372308  0.018274987
#> [25,] 0.06516334 0.07409195 0.1130824 0.1849187 0.4049261  0.019010862
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.6649345 0.7635493 0.7038882 0.6045379 0.6476671 0.6746493 0.8652795
#>  [2,] 0.2636717 0.3722361 0.5412604 0.6883117 0.7632717 0.8040487 0.8656574
#>  [3,] 0.8995012 0.8050929 0.7159419 0.7612473 0.7510554 0.5006047 0.2953567
#>  [4,] 0.2970380 0.3233358 0.4546674 0.5391400 0.5235430 0.5378344 0.5664050
#>  [5,] 0.8918241 0.5290922 0.6418935 0.8423264 0.6730725 0.7835030 0.8808485
#>  [6,] 0.7445997 0.7690637 0.8372220 0.9582780 1.0169786 1.0133640 1.0078146
#>  [7,] 0.7139324 0.7921972 0.8654862 0.8956056 0.9507806 0.8682090 0.7931224
#>  [8,] 0.9447146 1.0628363 0.4528551 0.0890289 0.1834185 0.1265594 0.3191959
#>  [9,] 0.9546388 0.9199718 1.0009270 0.9439387 0.8226185 0.6490474 0.5718239
#> [10,] 0.9171080 0.8630916 0.7651389 0.7732347 0.6664285 0.5036175 0.5023535
#> [11,] 0.4110654 0.5340691 0.4724151 0.4526185 0.5991469 0.5157009 0.4753352
#> [12,] 0.6896699 0.7893314 0.7854482 0.8351631 0.9543611 0.9532411 0.9556094
#> [13,] 0.5759080 0.7317150 0.8990144 0.9393046 0.9398253 0.9629606 0.8952485
#> [14,] 0.4525008 0.5711201 0.7335772 0.7924749 0.8609647 0.9387952 1.0037536
#> [15,] 1.0451614 0.9247026 0.8012290 0.6446020 0.5189345 0.4767198 0.8033218
#> [16,] 0.5421398 0.6651330 0.7937031 0.9439053 0.9313230 0.9556523 0.9896649
#> [17,] 0.3023142 0.5069634 0.7456993 0.8503089 0.9935504 0.8563041 0.6848025
#> [18,] 0.4424053 0.4751562 0.4452572 0.3017871 0.3652311 0.4339831 0.5040717
#> [19,] 0.4309861 0.5693592 0.6829747 0.8418342 0.9008706 0.9046278 0.9800270
#> [20,] 0.6825539 0.4340501 0.2452184 0.3964461 0.8004609 0.8218114 0.4644529
#> [21,] 0.3761566 0.4145278 0.4829942 0.5043447 0.5620435 0.4725226 0.5496257
#> [22,] 0.2527509 0.3108491 0.3504675 0.5314224 0.7196795 0.8511519 0.8826674
#> [23,] 0.4375972 0.5483595 0.6319080 0.6788490 0.7504119 0.7417283 0.5690512
#> [24,] 0.2029959 0.3013060 0.5142825 0.7621352 0.8238037 0.8079804 0.8323414
#> [25,] 0.3899848 0.2473973 0.2672395 0.4124170 0.5695828 0.5430994 0.4808104
#>             V27       V28       V29          V3       V30       V31        V32
#>  [1,] 0.9948988 0.9497264 0.8443218 0.062208325 0.6336775 0.5570942 0.68965864
#>  [2,] 0.9385228 0.9691039 0.8279696 0.043246281 0.6663977 0.6385813 0.53927561
#>  [3,] 0.2975480 0.2301170 0.3345101 0.091481163 0.2806479 0.1618578 0.27530087
#>  [4,] 0.6436665 0.8287771 0.8739315 0.080194004 0.8457656 0.8830969 0.81480036
#>  [5,] 0.8352116 0.6970921 0.3393949 0.073895177 0.6453726 0.7153032 0.82167581
#>  [6,] 1.0054890 1.0169022 0.8853361 0.025054688 0.6537339 0.4485750 0.34913884
#>  [7,] 0.7560386 0.6974923 0.5333071 0.045358607 0.3972526 0.3745050 0.39160114
#>  [8,] 0.2095694 0.3559335 0.4654072 0.091264654 0.6993185 0.2661240 0.29680017
#>  [9,] 0.5909178 0.5470901 0.4717460 0.110417170 0.3215815 0.1231205 0.09805652
#> [10,] 0.4870903 0.4538134 0.2969723 0.076641658 0.2257892 0.1536310 0.15476867
#> [11,] 0.7187817 0.8879236 0.8763634 0.050463824 0.9176644 0.8185587 0.62251268
#> [12,] 0.9792130 0.9751373 0.8585487 0.038367855 0.6220099 0.3899480 0.28867128
#> [13,] 0.8033558 0.6263865 0.3577107 0.026388755 0.2966482 0.2888239 0.39642853
#> [14,] 0.9547997 0.7753152 0.7081218 0.019787775 0.8166017 0.7536930 0.46317519
#> [15,] 1.0204727 0.7701861 0.4704950 0.040088028 0.1329354 0.2743572 0.37400223
#> [16,] 0.9239460 0.8286363 0.7555032 0.013899810 0.6713073 0.6355230 0.59092179
#> [17,] 0.7006539 0.5827209 0.4673854 0.099116911 0.5189472 0.4360806 0.15550925
#> [18,] 0.7164857 0.9058769 0.9541153 0.013681824 0.8551831 0.8011417 0.75878885
#> [19,] 0.9692795 0.8713566 0.8497938 0.007226026 0.7569230 0.7266397 0.67523664
#> [20,] 0.3190204 0.6584162 0.8468220 0.064924089 0.5723573 0.6368672 0.75756768
#> [21,] 0.6327411 0.7564737 0.9210378 0.020555328 0.8865790 0.7578385 0.67682298
#> [22,] 0.7923598 0.6887408 0.5204865 0.013702460 0.3650168 0.3070140 0.23833212
#> [23,] 0.6485074 0.7529036 0.5670626 0.051939442 0.5843034 0.5779482 0.41227517
#> [24,] 0.7949553 0.8519925 0.9351601 0.027645487 0.8137906 0.6804459 0.56563027
#> [25,] 0.3781889 0.6082119 0.9313816 0.020703665 0.9843877 0.9399011 0.86765940
#>             V33        V34        V35        V36        V37        V38
#>  [1,] 0.5785650 0.29670375 0.03545322 0.28498768 0.33479658 0.41917877
#>  [2,] 0.4368598 0.25100697 0.13368051 0.10303952 0.21116140 0.19109881
#>  [3,] 0.3248241 0.36171137 0.43778324 0.32348598 0.34646446 0.42659619
#>  [4,] 0.6971699 0.58188072 0.55938263 0.39006065 0.18227233 0.25816278
#>  [5,] 0.6666194 0.52840812 0.29429991 0.06644852 0.08940179 0.38575704
#>  [6,] 0.2263978 0.07842171 0.11768651 0.18575524 0.23796973 0.33621477
#>  [7,] 0.6161782 0.70027910 0.60449578 0.51243636 0.45362715 0.34457671
#>  [8,] 0.5450271 0.73764315 0.88150628 0.86502562 0.77628699 0.89783286
#>  [9,] 0.2936788 0.28405887 0.25999630 0.25133886 0.35991965 0.37359145
#> [10,] 0.1033491 0.10244376 0.20633315 0.04588089 0.08647287 0.19207023
#> [11,] 0.2525028 0.29438378 0.36193663 0.24566446 0.29815801 0.34034870
#> [12,] 0.2144128 0.19263345 0.18087031 0.09001149 0.06579923 0.09774925
#> [13,] 0.4047175 0.37263847 0.21098488 0.25257629 0.31137156 0.29047639
#> [14,] 0.1846180 0.14179895 0.14630025 0.15891960 0.24889846 0.17515331
#> [15,] 0.3162914 0.36739519 0.38632955 0.23460672 0.29193815 0.35806612
#> [16,] 0.5515695 0.53964492 0.51210014 0.46417127 0.41176991 0.30320644
#> [17,] 0.1749053 0.07650818 0.41227159 0.70407745 0.71794246 0.34023916
#> [18,] 0.7363237 0.60749524 0.57073912 0.59193002 0.54621184 0.38157594
#> [19,] 0.5965447 0.60898167 0.60901506 0.52109900 0.46316678 0.38985016
#> [20,] 0.6424820 0.89891542 0.81352331 0.74859185 0.81260396 0.59337892
#> [21,] 0.6774377 0.58127336 0.47675509 0.45862044 0.42266910 0.40801143
#> [22,] 0.2066955 0.25416608 0.28862594 0.27748254 0.20709779 0.15883589
#> [23,] 0.2581069 0.23904545 0.48900061 0.57253147 0.56295261 0.42256299
#> [24,] 0.6111914 0.72002110 0.67619386 0.51652395 0.26027547 0.15435445
#> [25,] 0.7796110 0.69916074 0.59614589 0.48840665 0.44623269 0.16775338
#>              V39           V4        V40        V41        V42        V43
#>  [1,] 0.59406945  0.041527450 0.45072553 0.34343528 0.52200834 0.49539107
#>  [2,] 0.18212316  0.057846781 0.19329384 0.25051978 0.25015990 0.22704367
#>  [3,] 0.33185869  0.113033949 0.29807572 0.26044611 0.26941241 0.34130181
#>  [4,] 0.20717562  0.085644421 0.17882657 0.51469922 0.60371639 0.44716881
#>  [5,] 0.59624445  0.114787097 0.34243360 0.38863856 0.48046101 0.46258321
#>  [6,] 0.32535520  0.051406587 0.31667772 0.34733405 0.33391201 0.23552865
#>  [7,] 0.37870353  0.061131100 0.38798402 0.28632546 0.25725897 0.25564481
#>  [8,] 0.80442915  0.076216149 0.53564056 0.30281534 0.14830932 0.13166808
#>  [9,] 0.35259514  0.085319110 0.41403774 0.47551999 0.56053704 0.32962790
#> [10,] 0.22094286  0.061432698 0.18217546 0.13781538 0.19195636 0.20556154
#> [11,] 0.26686211  0.024292932 0.18153578 0.60748585 0.65947001 0.60516723
#> [12,] 0.11611081  0.024303908 0.11816559 0.06380957 0.11348563 0.19470822
#> [13,] 0.38078211  0.028837526 0.38350016 0.24750663 0.12371381 0.29562120
#> [14,] 0.10284626  0.033110297 0.07854413 0.07710104 0.19652251 0.23459734
#> [15,] 0.25094187  0.021036590 0.29587241 0.22212458 0.15986471 0.17968470
#> [16,] 0.11955335  0.044086258 0.18497188 0.18261746 0.09762615 0.06465889
#> [17,] 0.08798832  0.082652275 0.31744664 0.42959927 0.39296379 0.24802692
#> [18,] 0.16217039 -0.001775884 0.04583986 0.01826728 0.15344477 0.23119648
#> [19,] 0.22446847  0.005260942 0.19257849 0.18536280 0.04740550 0.03714045
#> [20,] 0.44741414  0.069852522 0.39157735 0.30620880 0.29224786 0.09198601
#> [21,] 0.41122018  0.024533539 0.34342729 0.25806178 0.13638093 0.09985104
#> [22,] 0.19959287  0.013715246 0.20809755 0.22340300 0.18599114 0.15496783
#> [23,] 0.25916397  0.035631484 0.31968273 0.31202799 0.28291450 0.24684942
#> [24,] 0.15805709  0.039759596 0.22068843 0.20505753 0.23283745 0.30428643
#> [25,] 0.04327591  0.035576024 0.12151005 0.15982271 0.16638104 0.22171752
#>               V44         V45        V46        V47        V48         V49
#>  [1,] 0.513554616 0.539093197 0.24579004 0.17660123 0.20970497 0.109504205
#>  [2,] 0.158225188 0.125665416 0.12296173 0.14842260 0.15171783 0.081001250
#>  [3,] 0.258520084 0.101923745 0.02590294 0.07438385 0.06270748 0.024925334
#>  [4,] 0.512254108 0.600881240 0.54712425 0.35426500 0.18025097 0.082450120
#>  [5,] 0.418047020 0.545756906 0.56831560 0.32042452 0.24995457 0.130958065
#>  [6,] 0.199030387 0.161015068 0.11691057 0.09890192 0.08315758 0.048277695
#>  [7,] 0.202610824 0.148805877 0.08943897 0.09887822 0.12060895 0.057968009
#>  [8,] 0.252657873 0.459354157 0.34546122 0.17545089 0.19175106 0.096650561
#>  [9,] 0.167229912 0.144668267 0.17669970 0.08291243 0.04094667 0.034022077
#> [10,] 0.205836187 0.128516583 0.07023660 0.07224412 0.08541396 0.072335856
#> [11,] 0.457730199 0.543365169 0.61161349 0.45254524 0.28141183 0.187521445
#> [12,] 0.215550918 0.207237603 0.22698909 0.18605036 0.11812192 0.058632725
#> [13,] 0.323644801 0.341434964 0.28022315 0.16482755 0.07834474 0.051060366
#> [14,] 0.180421424 0.142270352 0.13742928 0.10316422 0.02012730 0.021158640
#> [15,] 0.106085187 0.113496808 0.12826152 0.12449731 0.09215437 0.057762340
#> [16,] 0.006522688 0.020372070 0.01001775 0.01389371 0.01950552 0.007695783
#> [17,] 0.193856521 0.225030522 0.19471118 0.18965403 0.17193671 0.123315029
#> [18,] 0.247717899 0.200917511 0.15301062 0.09360971 0.11365334 0.063722648
#> [19,] 0.020955572 0.007175247 0.01767796 0.01861550 0.03516070 0.010867415
#> [20,] 0.193554499 0.154768932 0.15148942 0.11381265 0.06289333 0.040679800
#> [21,] 0.085913099 0.091527843 0.03152777 0.02226691 0.03052861 0.013444792
#> [22,] 0.121935264 0.092715679 0.08995888 0.03957285 0.02830327 0.023908968
#> [23,] 0.191546629 0.195733944 0.16193252 0.14559959 0.10735812 0.069393797
#> [24,] 0.186889940 0.065470267 0.05774788 0.05563027 0.04331389 0.012081826
#> [25,] 0.342449682 0.328211790 0.21344865 0.09751099 0.01410712 0.009856524
#>                V5           V50         V51         V52         V53         V54
#>  [1,] 0.079208552  0.0247197243 0.035094031 0.019687466 0.003862145 0.011173568
#>  [2,] 0.066842156  0.0159003845 0.019254171 0.014694465 0.008098275 0.010945591
#>  [3,] 0.114745889 -0.0005552771 0.002239622 0.005504264 0.015899845 0.016884568
#>  [4,] 0.074523876  0.0310559152 0.042696319 0.044984277 0.023615127 0.019104393
#>  [5,] 0.252098404  0.0834091139 0.024303424 0.021152375 0.036656426 0.024052375
#>  [6,] 0.109674461  0.0147450075 0.017417450 0.011355751 0.017510911 0.015753973
#>  [7,] 0.039968681  0.0200828644 0.012292734 0.012643061 0.008698384 0.017381909
#>  [8,] 0.039365896  0.0399616913 0.026778911 0.015393579 0.013136760 0.017804193
#>  [9,] 0.160156925  0.0117742325 0.014203564 0.017785178 0.020113946 0.017420851
#> [10,] 0.113451191  0.0244301493 0.034067992 0.018646454 0.023450080 0.017655053
#> [11,] 0.058300277  0.0777768342 0.074632799 0.034962942 0.015516398 0.026240521
#> [12,] 0.071254704  0.0175461738 0.016458260 0.014248169 0.005969350 0.003393792
#> [13,] 0.045067253  0.0212467000 0.012080411 0.005095474 0.009897726 0.012169975
#> [14,] 0.019850798  0.0136175767 0.015592457 0.013581649 0.011401418 0.006142481
#> [15,] 0.059460814  0.0361207824 0.005370655 0.013800340 0.014847583 0.014752565
#> [16,] 0.050780676  0.0108844972 0.009161091 0.007887836 0.004715030 0.002795752
#> [17,] 0.041629070  0.0316187060 0.017123784 0.010182837 0.005016567 0.012358038
#> [18,] 0.006890658  0.0131226672 0.011002605 0.008814001 0.002597004 0.007391911
#> [19,] 0.049660455  0.0073399039 0.004298565 0.002993293 0.011489347 0.010959110
#> [20,] 0.120526637  0.0102887823 0.014869694 0.008181835 0.015310806 0.004686755
#> [21,] 0.024931084  0.0092663153 0.003030551 0.004793971 0.008367198 0.007043780
#> [22,] 0.029908085  0.0102959067 0.007338707 0.012660293 0.004674776 0.005989965
#> [23,] 0.068084411  0.0250324964 0.010087781 0.015502110 0.010255671 0.010127396
#> [24,] 0.048681213  0.0102931199 0.012054899 0.007842085 0.007711872 0.006330735
#> [25,] 0.034701806  0.0163233853 0.012599152 0.001544056 0.006034612 0.011811098
#>                V55         V56         V57         V58         V59         V6
#>  [1,] 0.0141561262 0.004030427 0.006404929 0.002918927 0.003612308 0.09681692
#>  [2,] 0.0069479224 0.008913576 0.006855333 0.008487991 0.008326009 0.12070647
#>  [3,] 0.0066791341 0.020539577 0.005546331 0.005348010 0.019098729 0.08290903
#>  [4,] 0.0142955676 0.008686867 0.015235391 0.011433944 0.014473322 0.10660389
#>  [5,] 0.0453197725 0.039949577 0.036022184 0.044677983 0.024647552 0.12722204
#>  [6,] 0.0102674787 0.011828359 0.006929546 0.008586867 0.006565025 0.14494692
#>  [7,] 0.0116066046 0.009689162 0.005783045 0.014545664 0.012671780 0.05246917
#>  [8,] 0.0161796556 0.018037575 0.020145419 0.013571221 0.010117494 0.11229323
#>  [9,] 0.0008390561 0.007816510 0.007658089 0.006817065 0.006316945 0.19088321
#> [10,] 0.0275049363 0.005983126 0.003440340 0.021552102 0.014644827 0.15552412
#> [11,] 0.0121190582 0.015148180 0.009467591 0.021889781 0.009541831 0.10089971
#> [12,] 0.0041020613 0.005171781 0.003587902 0.001866058 0.001834036 0.09986430
#> [13,] 0.0114263792 0.010680816 0.007750134 0.008811934 0.007651208 0.08005627
#> [14,] 0.0037853551 0.005521806 0.003792398 0.004347391 0.005693482 0.05197341
#> [15,] 0.0079204563 0.006280831 0.006408967 0.008551597 0.004461204 0.05141093
#> [16,] 0.0022696143 0.006308991 0.004822021 0.002346992 0.002727567 0.08174233
#> [17,] 0.0066263427 0.005335522 0.010989437 0.006950612 0.007883451 0.06924189
#> [18,] 0.0090462149 0.002390157 0.002199919 0.004081228 0.008376467 0.10235818
#> [19,] 0.0075479429 0.008243182 0.003510891 0.003253499 0.001496473 0.09949914
#> [20,] 0.0177107826 0.005990584 0.009861010 0.006554294 0.012279155 0.19974130
#> [21,] 0.0051462675 0.005515220 0.005034373 0.003144076 0.006490290 0.06083449
#> [22,] 0.0049993332 0.002778206 0.002493567 0.004385033 0.002864927 0.04785403
#> [23,] 0.0081046121 0.006824794 0.016757015 0.007922452 0.006366900 0.13463815
#> [24,] 0.0043582198 0.005869427 0.010426048 0.008692316 0.007712695 0.07344846
#> [25,] 0.0016992961 0.002143511 0.006210951 0.005692884 0.007576837 0.02351618
#>                V60         V7           V8         V9
#>  [1,] 0.0024058806 0.09560588  0.155320266 0.25293437
#>  [2,] 0.0058292716 0.13235155  0.142281826 0.17833783
#>  [3,] 0.0083688956 0.05409485  0.094808347 0.01184259
#>  [4,] 0.0051529783 0.05752910  0.104675320 0.23277392
#>  [5,] 0.0097674705 0.18089674  0.098408633 0.24918789
#>  [6,] 0.0031564482 0.13407106  0.161144679 0.20160698
#>  [7,] 0.0063554876 0.14192981  0.286330336 0.44784265
#>  [8,] 0.0100068151 0.21992861  0.315659371 0.55114795
#>  [9,] 0.0069832636 0.11585844  0.166752350 0.15751912
#> [10,] 0.0139491296 0.18472308  0.121541689 0.19262422
#> [11,] 0.0081417028 0.16152834  0.219862012 0.26817486
#> [12,] 0.0060110349 0.13521189  0.158589189 0.21799658
#> [13,] 0.0060172401 0.08535672  0.120291672 0.24941783
#> [14,] 0.0104410269 0.08018363  0.136239286 0.18818958
#> [15,] 0.0017856041 0.06137966 -0.006726871 0.06519773
#> [16,] 0.0032219013 0.10021060  0.112798338 0.16756721
#> [17,] 0.0043281847 0.14985586  0.213519621 0.23178549
#> [18,] 0.0043754194 0.09159761  0.016459512 0.04349399
#> [19,] 0.0024287882 0.10070447  0.075832728 0.16368241
#> [20,] 0.0005270129 0.19005032  0.197140788 0.17594476
#> [21,] 0.0053515137 0.04385936  0.081842280 0.07117135
#> [22,] 0.0028752312 0.05654622  0.061943579 0.08517221
#> [23,] 0.0050544834 0.21145625  0.287824187 0.33591632
#> [24,] 0.0066504054 0.06491958  0.090420505 0.08052576
#> [25,] 0.0010483472 0.08361899  0.080446408 0.05766950
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551