Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13       V14
#>  [1,] 0.031335771 0.430387840 0.51721391 0.52125539 0.49468454 0.3765895
#>  [2,] 0.026155309 0.396917226 0.40061841 0.33885270 0.29240651 0.1602605
#>  [3,] 0.032919232 0.218440406 0.26703299 0.33567244 0.34134724 0.3249658
#>  [4,] 0.028199664 0.251269100 0.29477972 0.32441784 0.27518744 0.2208345
#>  [5,] 0.018206345 0.088064752 0.15685685 0.18302940 0.16633360 0.1252570
#>  [6,] 0.045440948 0.154414093 0.24645017 0.25292724 0.36105627 0.3409835
#>  [7,] 0.033676778 0.222356834 0.31349275 0.34619149 0.29937396 0.1346069
#>  [8,] 0.007671372 0.339647599 0.31581151 0.31697750 0.22936772 0.1374390
#>  [9,] 0.026355987 0.267348313 0.31801782 0.35874770 0.45806456 0.5779748
#> [10,] 0.020900000 0.391400000 0.35040000 0.36690000 0.39430000 0.3311000
#> [11,] 0.010002896 0.317222851 0.31329237 0.33612833 0.25435561 0.1869495
#> [12,] 0.042234249 0.179090060 0.23609577 0.31627843 0.35658181 0.4191831
#> [13,] 0.102133858 0.554621546 0.41109636 0.17206546 0.17443809 0.2915595
#> [14,] 0.030918838 0.247076023 0.32851140 0.35149766 0.29249880 0.2492054
#> [15,] 0.020718399 0.093743645 0.09948198 0.13430466 0.13966025 0.1365216
#> [16,] 0.023050922 0.226927842 0.23558691 0.12501725 0.21440762 0.4976811
#> [17,] 0.009631029 0.105206096 0.08767687 0.13110327 0.26733186 0.2374871
#> [18,] 0.030389687 0.006800711 0.10468333 0.32262075 0.30458638 0.2712118
#> [19,] 0.029995971 0.086357642 0.24805457 0.36658093 0.27973446 0.2808161
#> [20,] 0.024093327 0.395908171 0.29344687 0.12710274 0.19140967 0.4738365
#> [21,] 0.027000000 0.152000000 0.16750000 0.13700000 0.13610000 0.1345000
#> [22,] 0.012055747 0.130269169 0.10720272 0.10743624 0.09323128 0.1081011
#> [23,] 0.014003625 0.097071560 0.12381105 0.09641391 0.07773815 0.1299726
#> [24,] 0.017277875 0.190671923 0.24831492 0.22455388 0.29683234 0.4239060
#> [25,] 0.026583980 0.087800600 0.13593911 0.25310605 0.21337277 0.2000389
#>              V15        V16       V17       V18       V19          V2
#>  [1,] 0.27665787 0.27039328 0.2968997 0.3237637 0.5195332 0.042357521
#>  [2,] 0.17121796 0.20608684 0.2259634 0.2795338 0.2427483 0.038744178
#>  [3,] 0.39370547 0.48958725 0.5370167 0.6203653 0.6686601 0.036032301
#>  [4,] 0.15805465 0.18972601 0.2234388 0.3573879 0.5849638 0.047259312
#>  [5,] 0.06680098 0.12995712 0.1822386 0.3090371 0.3689900 0.042300406
#>  [6,] 0.25977813 0.30059294 0.3664394 0.2635535 0.3432040 0.078244937
#>  [7,] 0.11373952 0.14648537 0.1834944 0.2352636 0.2090624 0.009851045
#>  [8,] 0.19848350 0.22308854 0.1909130 0.1745102 0.1286864 0.011166792
#>  [9,] 0.60518223 0.62074187 0.6618362 0.7582126 0.8628779 0.044443818
#> [10,] 0.33310000 0.30020000 0.2324000 0.1381000 0.3450000 0.019100000
#> [11,] 0.10731483 0.21464905 0.2729955 0.3277250 0.2350241 0.036044526
#> [12,] 0.51781821 0.57763007 0.6894377 0.7674084 0.8478049 0.063232898
#> [13,] 0.46646442 0.51341996 0.2530434 0.2021998 0.3201679 0.093903030
#> [14,] 0.21000408 0.18799701 0.2413851 0.2888917 0.3751302 0.046895059
#> [15,] 0.10892045 0.11785553 0.1443756 0.1950128 0.2832161 0.024021644
#> [16,] 0.70352594 0.76749092 0.8012633 0.6305966 0.4260041 0.067308832
#> [17,] 0.18371848 0.27973621 0.3034216 0.4429314 0.6462160 0.025764118
#> [18,] 0.06616936 0.08361303 0.1857849 0.4595348 0.3728152 0.032223970
#> [19,] 0.32108167 0.43103925 0.5658648 0.4570314 0.4366142 0.061887901
#> [20,] 0.60561989 0.64920829 0.7245186 0.6163804 0.3313343 0.029761124
#> [21,] 0.21440000 0.53540000 0.6830000 0.5600000 0.3093000 0.009200000
#> [22,] 0.06775811 0.11864192 0.2539830 0.2874563 0.3562824 0.013094964
#> [23,] 0.15149765 0.17983570 0.1847249 0.2592081 0.3148817 0.014037191
#> [24,] 0.54614495 0.72992944 0.8813000 0.9063321 0.7924777 0.018519885
#> [25,] 0.21742629 0.36919542 0.6762679 0.6224616 0.3588921 0.035231308
#>              V20        V21       V22       V23       V24       V25       V26
#>  [1,] 0.71410825 0.80670119 0.8619239 0.8006265 0.8137026 0.8184454 0.8571040
#>  [2,] 0.18110477 0.30816957 0.6626244 0.8388922 0.8538033 0.8559556 0.8595753
#>  [3,] 0.76555905 0.80813377 0.8662595 0.9430973 0.9936796 0.9945425 0.9857141
#>  [4,] 0.71924302 0.82848043 0.8226156 0.8289655 0.9500986 0.9435427 0.9671176
#>  [5,] 0.55023037 0.61487945 0.6908237 0.7160392 0.5894157 0.6361191 0.8033962
#>  [6,] 0.44374785 0.57344586 0.6934199 0.6898284 0.6645542 0.5758627 0.5567799
#>  [7,] 0.23947907 0.32505110 0.5937107 0.8355525 0.9049394 0.9846054 0.9933836
#>  [8,] 0.08791866 0.14257527 0.3162851 0.5205474 0.6688177 0.7853191 0.8940957
#>  [9,] 0.89334971 0.89637608 0.9318046 0.9090564 0.8231844 0.7282358 0.7115369
#> [10,] 0.44280000 0.48900000 0.3677000 0.4379000 0.4864000 0.6207000 0.7256000
#> [11,] 0.13189781 0.05514795 0.1300266 0.3379672 0.4766521 0.6214492 0.8117737
#> [12,] 0.88350035 0.88956650 0.8527060 0.7589904 0.6042250 0.4303564 0.2995710
#> [13,] 0.86085960 0.97422980 0.4151466 0.1010337 0.1724522 0.2130177 0.3472600
#> [14,] 0.38822586 0.47670077 0.5268933 0.6234914 0.6866215 0.6385057 0.6722366
#> [15,] 0.28798575 0.27276583 0.2803548 0.3711979 0.4761798 0.4880855 0.5121203
#> [16,] 0.30745439 0.52577554 0.7563236 0.8527553 1.0000000 0.8530250 0.6867849
#> [17,] 0.76753266 0.78417867 0.6992861 0.4822882 0.4404518 0.6058916 0.7098541
#> [18,] 0.58479140 0.70691508 0.7471018 0.8046505 0.8062590 0.8785967 0.7415916
#> [19,] 0.65091256 0.39630523 0.2006415 0.3559880 0.9245732 0.9483976 0.4630869
#> [20,] 0.35276497 0.51460671 0.7541108 0.8484499 0.8370824 0.6835931 0.6543652
#> [21,] 0.32260000 0.44300000 0.5573000 0.5782000 0.6173000 0.8132000 0.9819000
#> [22,] 0.50172678 0.63159690 0.7959804 0.9165598 0.8873148 0.8538218 0.9259624
#> [23,] 0.34607981 0.41602416 0.5150839 0.6192970 0.6598711 0.6689398 0.7669308
#> [24,] 0.60136772 0.58939862 0.6799214 0.6936084 0.6936481 0.6953106 0.7034400
#> [25,] 0.25957660 0.34791707 0.4849984 0.5736681 0.6222594 0.7283894 0.7943427
#>             V27       V28       V29         V3       V30       V31       V32
#>  [1,] 0.9192931 0.8517284 0.6840525 0.04173182 0.4840163 0.4246103 0.4643405
#>  [2,] 0.9418952 1.0257307 0.8050300 0.04487764 0.5552054 0.6605863 0.6653305
#>  [3,] 0.9621104 0.9779854 0.8306490 0.01491071 0.6197554 0.4931790 0.4003874
#>  [4,] 0.9888124 0.9811740 0.8731792 0.04876207 0.6443665 0.4151853 0.3072245
#>  [5,] 0.9503409 0.9800434 0.7834423 0.02536684 0.5680492 0.3429789 0.1751573
#>  [6,] 0.5665007 0.7380596 0.8151779 0.07941083 0.8234649 0.8940488 0.8613744
#>  [7,] 0.9502936 1.0063031 0.9140183 0.03047255 0.7501308 0.7743295 0.7060591
#>  [8,] 0.9471890 0.9558342 0.9080557 0.01457645 0.9178179 0.8831661 0.8030234
#>  [9,] 0.7099646 0.6573888 0.6001664 0.04785020 0.4242737 0.2448048 0.1249287
#> [10,] 0.6624000 0.7689000 0.7981000 0.04110000 0.8577000 0.9273000 0.7009000
#> [11,] 0.9528607 0.9596436 0.9540819 0.02464239 1.0265251 0.9620587 0.8613457
#> [12,] 0.3021887 0.2685143 0.2674219 0.07107961 0.3288418 0.2231703 0.2466321
#> [13,] 0.2401969 0.3934561 0.4640462 0.05860882 0.6154076 0.2840739 0.3465007
#> [14,] 0.8519896 0.9493584 0.8855183 0.05050137 0.7488440 0.6744433 0.5897071
#> [15,] 0.5061536 0.6493426 0.8079199 0.02001737 0.8639162 0.8407274 0.7622545
#> [16,] 0.6980126 0.5643903 0.4572663 0.11170327 0.5022077 0.4200412 0.1395305
#> [17,] 0.8075455 0.8002589 0.5921218 0.01770647 0.3460179 0.3220858 0.3449022
#> [18,] 0.6142031 0.5763222 0.5499119 0.05391904 0.4303299 0.4786840 0.5487322
#> [19,] 0.2718835 0.6643244 0.8098388 0.06519395 0.4198828 0.5218729 0.7023074
#> [20,] 0.7490464 0.6769281 0.4707971 0.05715892 0.4775125 0.5301080 0.3596724
#> [21,] 0.9823000 0.9166000 0.7423000 0.01450000 0.7736000 0.8473000 0.7352000
#> [22,] 0.9956976 0.9538106 0.8976066 0.01761726 0.7695463 0.6641539 0.6354895
#> [23,] 0.7591702 0.7125692 0.6794354 0.02413717 0.6161616 0.5740290 0.5332603
#> [24,] 0.5893968 0.4525641 0.4477543 0.03353805 0.4630316 0.3610775 0.2963725
#> [25,] 0.7932672 0.7589273 0.8543082 0.04546162 0.8965321 0.8540673 0.7469960
#>              V33        V34        V35        V36       V37        V38
#>  [1,] 0.46427473 0.37320817 0.18324447 0.30149748 0.3170804 0.35789393
#>  [2,] 0.50089076 0.20167317 0.08435432 0.03571459 0.1642953 0.19481656
#>  [3,] 0.30531459 0.08374314 0.05579842 0.11892579 0.2672638 0.36054649
#>  [4,] 0.27050065 0.23300855 0.16810408 0.02890855 0.0448390 0.08282341
#>  [5,] 0.19099506 0.42390604 0.68111683 0.78713650 0.7932510 0.69363860
#>  [6,] 0.72737355 0.65937783 0.64106596 0.36526115 0.1802830 0.27216571
#>  [7,] 0.53564036 0.21964585 0.09303774 0.11894254 0.1109779 0.11962610
#>  [8,] 0.71356407 0.66735283 0.64372832 0.68437942 0.6204668 0.48837545
#>  [9,] 0.05071513 0.08226391 0.20500767 0.27087264 0.2105810 0.23334680
#> [10,] 0.48510000 0.34090000 0.14060000 0.11470000 0.1433000 0.18200000
#> [11,] 0.74679103 0.67768405 0.71834067 0.76260678 0.7489224 0.59137853
#> [12,] 0.33339804 0.31214497 0.31195084 0.23168455 0.2944676 0.38089569
#> [13,] 0.55201624 0.68813663 0.78259230 0.79052023 0.7759389 0.83615114
#> [14,] 0.54922259 0.41285645 0.20559129 0.12042608 0.2377470 0.20463412
#> [15,] 0.76063962 0.77437012 0.75805710 0.67095538 0.5490227 0.37004918
#> [16,] 0.16091596 0.05430891 0.38781389 0.70749505 0.7260640 0.35301345
#> [17,] 0.30642657 0.27818963 0.32913165 0.32632108 0.4034811 0.40005013
#> [18,] 0.58463428 0.53219753 0.59793489 0.63799288 0.5097055 0.32456198
#> [19,] 0.58148498 0.89882986 0.79775318 0.72404203 0.7748179 0.44473259
#> [20,] 0.24178693 0.12298810 0.38557677 0.60173968 0.6289935 0.27591135
#> [21,] 0.66710000 0.60830000 0.62390000 0.59720000 0.5715000 0.52420000
#> [22,] 0.60696765 0.64075558 0.66920960 0.61419756 0.5459640 0.44713019
#> [23,] 0.52773215 0.60866005 0.63977779 0.59699600 0.5006217 0.47871346
#> [24,] 0.27948122 0.23054050 0.28833182 0.36586773 0.3113422 0.19788583
#> [25,] 0.69131002 0.76667696 0.88944513 0.97011092 0.9131001 0.79380639
#>              V39         V4        V40       V41        V42        V43
#>  [1,] 0.47155976 0.03700386 0.41897028 0.2062563 0.30219050 0.35872539
#>  [2,] 0.18763019 0.07710093 0.10611352 0.1527882 0.05207826 0.08076751
#>  [3,] 0.33101580 0.05275770 0.33401086 0.3461856 0.34071719 0.26974998
#>  [4,] 0.09207561 0.03018665 0.12618946 0.0336443 0.10071158 0.21642796
#>  [5,] 0.61239217 0.02516215 0.43555097 0.3046264 0.24255356 0.32951342
#>  [6,] 0.27776780 0.07697079 0.31709168 0.6518620 0.70143984 0.42503354
#>  [7,] 0.17018713 0.05624391 0.02965467 0.1565257 0.08393901 0.11848198
#>  [8,] 0.40010643 0.05718846 0.42284525 0.4514878 0.43411628 0.43058196
#>  [9,] 0.32585084 0.05993041 0.29736649 0.2900821 0.28218931 0.22041326
#> [10,] 0.36050000 0.03210000 0.55290000 0.5988000 0.50770000 0.55120000
#> [11,] 0.48842340 0.06602923 0.51095216 0.4526475 0.39264003 0.48288982
#> [12,] 0.31139026 0.07052119 0.27577303 0.2606954 0.27901386 0.25899964
#> [13,] 0.74169973 0.07241545 0.54237409 0.2799407 0.17935430 0.20289798
#> [14,] 0.24635305 0.06114660 0.33379997 0.3556453 0.36763519 0.29714011
#> [15,] 0.27076652 0.02762315 0.30250013 0.3133780 0.29056929 0.26334791
#> [16,] 0.05945238 0.09673785 0.27709356 0.4143801 0.38725590 0.24255901
#> [17,] 0.32314459 0.02844068 0.27028722 0.2235200 0.15171546 0.09289368
#> [18,] 0.20373072 0.05107029 0.43735965 0.4771726 0.54668386 0.51533601
#> [19,] 0.30040704 0.09237880 0.31522115 0.2806919 0.28957372 0.04848461
#> [20,] 0.27780569 0.02854985 0.48058820 0.4493162 0.34600420 0.22649292
#> [21,] 0.29240000 0.02780000 0.15360000 0.2003000 0.20310000 0.22070000
#> [22,] 0.28192466 0.03876030 0.20945814 0.1727811 0.08692358 0.07268840
#> [23,] 0.43426119 0.02277977 0.38261360 0.2985547 0.25960770 0.24422200
#> [24,] 0.17067800 0.03553089 0.27364418 0.2942665 0.23711176 0.14007657
#> [25,] 0.64574738 0.03069046 0.46351900 0.3600068 0.28139141 0.29465753
#>              V44        V45         V46        V47        V48         V49
#>  [1,] 0.39911708 0.39447849 0.188881058 0.10246442 0.11228376 0.064805440
#>  [2,] 0.14242320 0.06827959 0.065796179 0.08205202 0.11596855 0.079110083
#>  [3,] 0.22154422 0.16834697 0.117140261 0.11020422 0.09622638 0.055387818
#>  [4,] 0.24806451 0.22579613 0.241783081 0.18443623 0.10982679 0.058021146
#>  [5,] 0.36214637 0.40666086 0.349246762 0.23119503 0.12909917 0.082904994
#>  [6,] 0.38664529 0.56047667 0.529179276 0.28636349 0.13728288 0.067029689
#>  [7,] 0.16905929 0.09337110 0.005736865 0.07725357 0.12557359 0.065312637
#>  [8,] 0.48697403 0.46036784 0.316361064 0.22204398 0.17710758 0.097507273
#>  [9,] 0.20579648 0.16009984 0.141115173 0.11229404 0.08176848 0.058305357
#> [10,] 0.50270000 0.70340000 0.590400000 0.40690000 0.27610000 0.158400000
#> [11,] 0.56402782 0.48256536 0.297411972 0.17570445 0.15748210 0.100617561
#> [12,] 0.16423598 0.09458980 0.065566078 0.06226237 0.03392129 0.027500083
#> [13,] 0.26201601 0.45908373 0.347229142 0.15373187 0.13597033 0.091627936
#> [14,] 0.16059273 0.16526792 0.169485443 0.18852917 0.16937764 0.082103391
#> [15,] 0.27607712 0.24644149 0.167531268 0.09935229 0.05219267 0.030778920
#> [16,] 0.20141697 0.23024369 0.184228996 0.18704184 0.17228550 0.128338322
#> [17,] 0.08518413 0.10047612 0.096591528 0.07879421 0.06139219 0.031501626
#> [18,] 0.33966742 0.20040008 0.141523989 0.07719621 0.05469625 0.028499818
#> [19,] 0.18657878 0.15427166 0.163061603 0.12603035 0.07422017 0.044818160
#> [20,] 0.13025214 0.13240593 0.129377641 0.12751405 0.11146031 0.071311691
#> [21,] 0.17780000 0.13530000 0.137300000 0.07490000 0.04720000 0.032500000
#> [22,] 0.06700943 0.03416624 0.031356934 0.02049619 0.02742079 0.009737562
#> [23,] 0.20792314 0.14600883 0.106834499 0.08359508 0.06188477 0.030561831
#> [24,] 0.14613927 0.13964704 0.099313871 0.10609310 0.09394962 0.048738366
#> [25,] 0.22867012 0.21364337 0.180472317 0.11304269 0.04959582 0.023572125
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.05322015 0.014237059 0.021039460 0.013013377 0.006306469 0.010598604
#>  [2,] 0.10381930 0.028760873 0.015996563 0.009408249 0.005344163 0.007207676
#>  [3,] 0.08348141 0.015296178 0.014806218 0.010273706 0.015895500 0.011588720
#>  [4,] 0.05831434 0.019118701 0.020654963 0.014739030 0.006409755 0.004747885
#>  [5,] 0.02910570 0.026824231 0.013286146 0.012432071 0.005660286 0.003232613
#>  [6,] 0.09437835 0.036863703 0.033335029 0.032361657 0.020636892 0.015263786
#>  [7,] 0.07717594 0.017589503 0.005164006 0.006221659 0.009418992 0.013452839
#>  [8,] 0.08690585 0.025297153 0.012787341 0.009443671 0.011698731 0.006195883
#>  [9,] 0.12556794 0.021975391 0.021884166 0.019119094 0.011151699 0.022267415
#> [10,] 0.06980000 0.051000000 0.005400000 0.007800000 0.020100000 0.010400000
#> [11,] 0.10386168 0.029373942 0.012342118 0.017122567 0.010431000 0.011143317
#> [12,] 0.09534362 0.021874712 0.013407037 0.018363778 0.016865543 0.016378529
#> [13,] 0.06463744 0.045181056 0.020358710 0.012471299 0.011926895 0.024625640
#> [14,] 0.07773141 0.015679917 0.023027688 0.017909104 0.010961550 0.013225515
#> [15,] 0.03760742 0.015721515 0.011457975 0.006269147 0.008443016 0.010320912
#> [16,] 0.04605854 0.033867923 0.016221321 0.009708281 0.004212627 0.013509519
#> [17,] 0.01388124 0.022873571 0.013114312 0.010264776 0.014816648 0.021252055
#> [18,] 0.04446826 0.021012761 0.003493198 0.007071708 0.013466440 0.011645341
#> [19,] 0.16351022 0.006220192 0.015547483 0.002776749 0.015241411 0.006920192
#> [20,] 0.02247445 0.014529934 0.022269081 0.006942955 0.007215514 0.009911657
#> [21,] 0.04120000 0.017900000 0.004500000 0.008400000 0.001000000 0.001800000
#> [22,] 0.05945676 0.009625177 0.005850611 0.006688594 0.006381522 0.003329424
#> [23,] 0.04221523 0.015095245 0.013320910 0.007338592 0.005732592 0.008703825
#> [24,] 0.09174770 0.011514700 0.015557358 0.009676860 0.007584686 0.008108878
#> [25,] 0.01996207 0.033494602 0.010152663 0.011500141 0.007210123 0.012219648
#>               V55         V56         V57         V58          V59         V6
#>  [1,] 0.009984900 0.006439799 0.006104642 0.007470137 0.0069202730 0.10424195
#>  [2,] 0.003886990 0.010448231 0.005767554 0.006563154 0.0092536568 0.12306697
#>  [3,] 0.007878975 0.012819371 0.005574099 0.005541651 0.0071927997 0.11859586
#>  [4,] 0.003324652 0.002372710 0.003579505 0.001274894 0.0022923483 0.06279324
#>  [5,] 0.006239723 0.004236813 0.004375773 0.003801649 0.0048877778 0.05374579
#>  [6,] 0.013933795 0.010507080 0.008661994 0.011440617 0.0126476005 0.16085725
#>  [7,] 0.006600440 0.006169253 0.003505520 0.003327900 0.0061402007 0.09630358
#>  [8,] 0.003984494 0.006954906 0.005591612 0.004753482 0.0008477873 0.12740569
#>  [9,] 0.014193713 0.007430292 0.009876551 0.013386878 0.0113957020 0.17992729
#> [10,] 0.003900000 0.003100000 0.006200000 0.008700000 0.0070000000 0.15790000
#> [11,] 0.002671784 0.008762686 0.007259243 0.004727808 0.0013972251 0.10161699
#> [12,] 0.010643944 0.011621714 0.009370427 0.011824418 0.0115438012 0.11564727
#> [13,] 0.020047311 0.016502743 0.013671116 0.016038849 0.0135694629 0.10962149
#> [14,] 0.008233821 0.009022875 0.008356128 0.010967537 0.0111022809 0.14081604
#> [15,] 0.006529661 0.004929382 0.006751664 0.005811406 0.0070527669 0.05384339
#> [16,] 0.006218212 0.005844099 0.008711802 0.006588611 0.0078086933 0.06844550
#> [17,] 0.014930376 0.011054604 0.008914114 0.012167130 0.0035559707 0.02159755
#> [18,] 0.007985068 0.013906732 0.011628438 0.004983575 0.0063571227 0.10387641
#> [19,] 0.021042415 0.007571716 0.015018171 0.004546444 0.0200202032 0.23227491
#> [20,] 0.008903633 0.005508110 0.011292980 0.007544052 0.0082385715 0.09005510
#> [21,] 0.006800000 0.003900000 0.012000000 0.013200000 0.0070000000 0.07570000
#> [22,] 0.001962910 0.003767656 0.004860238 0.004068209 0.0023242051 0.07959742
#> [23,] 0.007945135 0.008651922 0.006920023 0.003353604 0.0041242779 0.07831576
#> [24,] 0.006972066 0.007848093 0.006760221 0.007570544 0.0101266008 0.11410861
#> [25,] 0.009220489 0.003096400 0.008738694 0.007610039 0.0074743427 0.05925342
#>               V60         V7         V8         V9
#>  [1,] 0.005371653 0.12336719 0.19770432 0.32021299
#>  [2,] 0.011909592 0.15118711 0.23303239 0.31125182
#>  [3,] 0.003687264 0.12816079 0.13875823 0.17387681
#>  [4,] 0.003985351 0.10946615 0.13024925 0.19533028
#>  [5,] 0.004951496 0.04432383 0.03447684 0.04348878
#>  [6,] 0.009704569 0.11931759 0.14151699 0.22715595
#>  [7,] 0.006618556 0.10039376 0.10666659 0.11010649
#>  [8,] 0.003195729 0.16207688 0.14231331 0.29039746
#>  [9,] 0.008874283 0.19534660 0.21594189 0.24972989
#> [10,] 0.004200000 0.14380000 0.14020000 0.30480000
#> [11,] 0.003025130 0.15677341 0.20934667 0.31860579
#> [12,] 0.008776482 0.11417610 0.08245180 0.12192006
#> [13,] 0.007577177 0.16792528 0.38241252 0.54845280
#> [14,] 0.008447079 0.14748264 0.15078004 0.19321949
#> [15,] 0.003290727 0.08084100 0.05956479 0.05437365
#> [16,] 0.003456726 0.15892336 0.20430889 0.21417447
#> [17,] 0.004487491 0.05915413 0.04953690 0.10237470
#> [18,] 0.013114954 0.11211433 0.12899250 0.01542503
#> [19,] 0.007253533 0.21956981 0.20193831 0.14641114
#> [20,] 0.007802089 0.06660670 0.21716634 0.29315678
#> [21,] 0.008800000 0.10260000 0.11380000 0.07940000
#> [22,] 0.002611743 0.06701206 0.08711071 0.14130010
#> [23,] 0.005589264 0.09417887 0.09282344 0.09688822
#> [24,] 0.008520217 0.14087199 0.14811231 0.14475065
#> [25,] 0.008401775 0.12369545 0.12444444 0.09620836
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623