Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.005000000  0.22820000 0.25210000 0.34840000 0.33090000 0.26140000
#>  [2,] 0.024761420  0.05514475 0.10145506 0.12743205 0.14123304 0.20418281
#>  [3,] 0.057979733 -0.09354908 0.11200235 0.29930093 0.31015853 0.30008371
#>  [4,] 0.027004150  0.26447497 0.22466346 0.26276281 0.26152315 0.22388281
#>  [5,] 0.024138960  0.22159016 0.26655705 0.30842776 0.32025753 0.36837289
#>  [6,] 0.015967319  0.22531254 0.25577542 0.31962328 0.30961952 0.27369313
#>  [7,] 0.037328517  0.20289465 0.27561652 0.28744689 0.32940119 0.41325970
#>  [8,] 0.013650413  0.33015671 0.35959113 0.38154169 0.31395563 0.19821581
#>  [9,] 0.033669556  0.28858637 0.36059131 0.39111385 0.40829864 0.35389477
#> [10,] 0.009459469  0.17688255 0.20638099 0.26404320 0.36753343 0.46164471
#> [11,] 0.030599082  0.24086088 0.30546281 0.31761149 0.27998918 0.16900775
#> [12,] 0.033500000  0.26600000 0.31880000 0.35530000 0.31160000 0.19650000
#> [13,] 0.027225995  0.31348533 0.30255566 0.22440229 0.25500442 0.25907339
#> [14,] 0.012795472  0.12540025 0.16777717 0.17478746 0.21082799 0.26918908
#> [15,] 0.011918919  0.18293247 0.24855010 0.19828594 0.15878258 0.19332512
#> [16,] 0.029300000  0.08600000 0.04140000 0.04720000 0.08350000 0.09380000
#> [17,] 0.013300452  0.05452684 0.02612434 0.14994851 0.16373868 0.11797102
#> [18,] 0.033075031  0.30582471 0.23705008 0.22046714 0.23680708 0.12768052
#> [19,] 0.028260863  0.14367760 0.15916100 0.12974671 0.20657924 0.37541388
#> [20,] 0.006667767  0.09873577 0.10594074 0.06950295 0.04935319 0.08646979
#> [21,] 0.019694445  0.13434573 0.13405395 0.08825350 0.10549436 0.16625291
#> [22,] 0.008768619  0.06851076 0.11701850 0.14264656 0.24720568 0.24998849
#> [23,] 0.049396987  0.11794465 0.21842680 0.34845445 0.29797194 0.16019487
#> [24,] 0.019780031  0.28964680 0.28252599 0.38462325 0.39784790 0.28114874
#> [25,] 0.013954701  0.08712878 0.11900008 0.13690203 0.14363842 0.15240897
#>             V15       V16       V17        V18        V19           V2
#>  [1,] 0.1782000 0.2055000 0.2298000 0.35450000 0.62180000  0.001700000
#>  [2,] 0.2540044 0.4011486 0.5096195 0.68710166 0.87483814  0.017809659
#>  [3,] 0.4535195 0.4830665 0.6868819 0.74688256 0.87011532  0.054524114
#>  [4,] 0.1852218 0.2202503 0.2258738 0.19305340 0.19441174  0.043535618
#>  [5,] 0.4251886 0.4994886 0.5556030 0.64732862 0.72141801  0.014781547
#>  [6,] 0.2095378 0.2233104 0.2096990 0.29214179 0.49306971  0.018784302
#>  [7,] 0.5237389 0.6687583 0.8207774 0.82563764 0.84367360  0.049086492
#>  [8,] 0.1409609 0.2151276 0.2765139 0.45300524 0.64655419  0.038684046
#>  [9,] 0.2937285 0.3661831 0.3868625 0.39082250 0.50323695  0.046736412
#> [10,] 0.5596888 0.6713210 0.6835184 0.78464708 0.90212041  0.020381712
#> [11,] 0.1532405 0.1791438 0.2146124 0.25924038 0.26282504  0.028815746
#> [12,] 0.1780000 0.2794000 0.2870000 0.39690000 0.55990000  0.025800000
#> [13,] 0.1891094 0.1688417 0.2171493 0.35048338 0.52555338  0.049452206
#> [14,] 0.1910016 0.2025445 0.2628064 0.36033840 0.43320590  0.006929364
#> [15,] 0.2216217 0.2054245 0.2327341 0.05660257 0.43522769 -0.018959775
#> [16,] 0.1466000 0.0809000 0.1179000 0.21790000 0.33260000  0.064400000
#> [17,] 0.1314881 0.1152861 0.1205843 0.15002739 0.07268594  0.028836695
#> [18,] 0.3617515 0.3258351 0.2293711 0.22098876 0.25159783 -0.006395741
#> [19,] 0.5552195 0.7178474 0.8031364 0.79420550 0.76330035  0.028433644
#> [20,] 0.1487436 0.1384164 0.1106655 0.15981094 0.19768319  0.004817747
#> [21,] 0.1634865 0.2285882 0.3442235 0.33430906 0.34698888  0.002457105
#> [22,] 0.2618387 0.3258187 0.3425386 0.41672191 0.48814149  0.001342510
#> [23,] 0.2216981 0.3428647 0.4815297 0.59767469 0.59897335  0.049158113
#> [24,] 0.2891112 0.4362738 0.7254151 0.58410257 0.17664847  0.019768799
#> [25,] 0.1443853 0.1872459 0.2556264 0.27885282 0.28769462  0.016362646
#>              V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.72650000 0.8346000 0.8268000 0.8366000 0.9408000 0.9510000 0.9801000
#>  [2,] 0.94883609 0.9320934 0.9738773 1.0128875 0.9895428 0.9400601 0.8656935
#>  [3,] 1.01576146 0.9598183 0.9107086 0.9106009 1.0056577 0.9332606 0.6940710
#>  [4,] 0.19902408 0.1885216 0.3024427 0.4589057 0.5587950 0.6183280 0.6833028
#>  [5,] 0.80606329 0.8734418 0.9276638 0.9388487 0.9662069 0.9685109 0.9534953
#>  [6,] 0.61682069 0.7560452 0.7834723 0.7525912 0.8667763 0.9162393 0.9740489
#>  [7,] 0.86540544 0.8590490 0.7318601 0.5283863 0.3567968 0.2418397 0.2394989
#>  [8,] 0.71688372 0.8047851 0.7932894 0.8725834 0.9895421 0.9758746 0.9425340
#>  [9,] 0.67451530 0.7359184 0.7745840 0.7762323 0.7988024 0.8116544 0.9025969
#> [10,] 1.01645779 1.0856707 1.0705167 1.0146427 0.8678305 0.6950911 0.7780656
#> [11,] 0.27769080 0.3878108 0.6012973 0.7541129 0.8210885 0.8801555 0.9013366
#> [12,] 0.69360000 0.7969000 0.7452000 0.8203000 0.9261000 0.8810000 0.8814000
#> [13,] 0.77323888 0.7551247 0.4169152 0.2218666 0.2538084 0.4198836 0.7513982
#> [14,] 0.47267500 0.5614821 0.6148676 0.5907638 0.7077356 0.9696028 0.9980316
#> [15,] 0.69935096 0.4501602 0.4003055 1.0467144 0.9540042 0.4834390 0.6160524
#> [16,] 0.32580000 0.2111000 0.2302000 0.3361000 0.4259000 0.4609000 0.2606000
#> [17,] 0.09705144 0.1986349 0.1470225 0.0971082 0.1576991 0.1409069 0.2998797
#> [18,] 0.38812423 0.2554966 0.6532335 0.6009113 0.3064641 0.2192584 0.5772522
#> [19,] 0.73038717 0.8469650 0.8919468 0.8635976 0.8773860 0.8617832 0.8331763
#> [20,] 0.23038719 0.2998576 0.4133076 0.5772770 0.7681158 0.9191571 0.8896014
#> [21,] 0.44667328 0.5440198 0.6828136 0.7958448 0.8220201 0.8904457 0.9776404
#> [22,] 0.53999633 0.5642610 0.6009180 0.6780973 0.7588804 0.8528962 0.8785889
#> [23,] 0.42926988 0.3066153 0.2863600 0.3323711 0.3070233 0.2940414 0.3358347
#> [24,] 0.28163556 0.7322592 0.7243490 0.3836989 0.3057369 0.5639758 0.5479375
#> [25,] 0.33788829 0.4011014 0.5366389 0.6404704 0.7296088 0.7772682 0.7739262
#>             V27       V28       V29            V3       V30       V31       V32
#>  [1,] 0.9974000 1.0000000 0.9036000  0.0270000000 0.6409000 0.3857000 0.2908000
#>  [2,] 0.7509626 0.5980868 0.4810214  0.0635065750 0.4256294 0.3134966 0.1854095
#>  [3,] 0.4671439 0.3592252 0.3483073  0.0920676805 0.3341250 0.2489129 0.2347540
#>  [4,] 0.7956917 0.8853860 0.8817810  0.0478521514 0.9021912 0.8798092 0.8126307
#>  [5,] 0.9267405 0.8369996 0.7200430  0.0219977371 0.5759325 0.4104278 0.3307828
#>  [6,] 1.0000000 0.9593270 0.8590597  0.0331856131 0.6636928 0.4748716 0.3312920
#>  [7,] 0.2886980 0.2484864 0.2516256  0.0467242775 0.3123112 0.2152079 0.3206078
#>  [8,] 0.9451784 0.9335142 0.8320137  0.0336019766 0.6158557 0.3749305 0.2820125
#>  [9,] 0.9346927 0.8937530 0.7387333  0.0471950680 0.5915116 0.5266964 0.5679750
#> [10,] 0.7119066 0.6152015 0.4797982  0.0323796215 0.3318915 0.3152080 0.2511018
#> [11,] 0.9198272 0.9516244 0.8617633  0.0395822067 0.7031593 0.6842191 0.5705095
#> [12,] 0.9301000 0.9955000 0.8576000  0.0398000000 0.6069000 0.3934000 0.2464000
#> [13,] 0.8609466 0.8345101 0.7748735  0.0471294253 0.7192811 0.4480946 0.4166428
#> [14,] 1.0078981 0.9592724 0.8516064  0.0064145890 0.7010108 0.5903493 0.4435989
#> [15,] 0.7141958 0.7402169 0.5468345  0.0306639767 0.3581282 0.2473150 0.1511302
#> [16,] 0.0874000 0.2862000 0.5606000  0.0390000000 0.8344000 0.8096000 0.7250000
#> [17,] 0.4833561 0.6038188 0.6286342  0.0107177868 0.6352052 0.7344569 0.7026255
#> [18,] 0.8659048 0.6966997 0.7545909  0.0009821567 0.9371076 0.6273283 0.3511134
#> [19,] 0.7592627 0.5758604 0.4122627  0.0433996092 0.4464335 0.4584852 0.3469865
#> [20,] 0.7732821 0.6729147 0.5418615 -0.0095324207 0.3725992 0.2340616 0.2029151
#> [21,] 1.0107503 0.9521308 0.8339306  0.0058985773 0.7533611 0.7280675 0.6347057
#> [22,] 0.8838795 0.8467286 0.7124916  0.0093717985 0.5621191 0.4874603 0.3992730
#> [23,] 0.3434084 0.4016630 0.6270403  0.0799319119 0.7199432 0.8351905 0.7909619
#> [24,] 0.5445815 0.9229618 0.8526675 -0.0018923064 0.4253718 0.2675714 0.2392997
#> [25,] 0.7028771 0.7386416 0.8125858  0.0318354144 0.7399115 0.6437398 0.6143597
#>              V33        V34        V35        V36        V37        V38
#>  [1,] 0.20400000 0.16530000 0.17690000 0.11400000 0.07400000 0.09410000
#>  [2,] 0.15443520 0.25856304 0.36199656 0.35785597 0.25992774 0.25109881
#>  [3,] 0.40883182 0.53462845 0.47496773 0.27427577 0.27724273 0.31575588
#>  [4,] 0.68258030 0.58755901 0.58285875 0.55567457 0.51667549 0.47579070
#>  [5,] 0.23102926 0.09714792 0.09825146 0.12874416 0.21444039 0.27215682
#>  [6,] 0.31040443 0.20151233 0.10519694 0.03841788 0.07905677 0.10758466
#>  [7,] 0.38865290 0.31987185 0.32040928 0.23288210 0.23178872 0.26788096
#>  [8,] 0.20555609 0.22187571 0.22011430 0.13999497 0.05747650 0.07286012
#>  [9,] 0.45997940 0.31320529 0.14733420 0.21068163 0.24255141 0.36046660
#> [10,] 0.05308405 0.08176198 0.19750018 0.29387956 0.24940630 0.14738521
#> [11,] 0.42417718 0.21562463 0.11111159 0.12652792 0.18293722 0.15480554
#> [12,] 0.16450000 0.11400000 0.09560000 0.00800000 0.07020000 0.09360000
#> [13,] 0.62375478 0.64866998 0.58629534 0.50289653 0.49730918 0.51117316
#> [14,] 0.29032609 0.16105350 0.07263330 0.27018603 0.26124302 0.32838646
#> [15,] 0.11318339 0.35444562 0.29238971 0.30052592 0.26292170 0.22071905
#> [16,] 0.80480000 0.94350000 1.00000000 0.89600000 0.55160000 0.30370000
#> [17,] 0.72620649 0.86606826 0.99656929 0.97882867 0.84620513 0.68875933
#> [18,] 0.72570583 0.92721848 0.80483263 0.76025419 0.57106202 0.43905307
#> [19,] 0.30905377 0.26980196 0.25654550 0.22938764 0.27986833 0.20380393
#> [20,] 0.21900664 0.25011687 0.20931750 0.18120288 0.17897465 0.17975827
#> [21,] 0.55027710 0.49285479 0.46925929 0.45678690 0.42340089 0.39342452
#> [22,] 0.31725417 0.14883539 0.17231639 0.27111148 0.20265365 0.16611159
#> [23,] 0.63550085 0.61615740 0.67289413 0.72065612 0.80720349 0.96356675
#> [24,] 0.08887696 0.14468033 0.28217933 0.16525510 0.20071546 0.21780212
#> [25,] 0.66398541 0.71334409 0.72656530 0.65766061 0.51015170 0.32569598
#>              V39           V4        V40         V41          V42        V43
#>  [1,] 0.06210000  0.045000000 0.04260000  0.05720000  0.106800000 0.19090000
#>  [2,] 0.35869788  0.128838547 0.35279242  0.37545573  0.471651388 0.35125536
#>  [3,] 0.42372397  0.041999038 0.45915254  0.35298076  0.361251615 0.30706670
#>  [4,] 0.41305526  0.058382658 0.41379327  0.51342809  0.477960091 0.42959879
#>  [5,] 0.29657967  0.063489447 0.30667349  0.28192208  0.273816541 0.24062632
#>  [6,] 0.11802299  0.037370906 0.13969143  0.06209716  0.146618505 0.24178236
#>  [7,] 0.19535414  0.063026462 0.09579924  0.08055204  0.199818278 0.19904541
#>  [8,] 0.15985538  0.010308769 0.15662537  0.08593615  0.115854155 0.17592933
#>  [9,] 0.48898211  0.050033474 0.38195985  0.31186951  0.415354228 0.40574939
#> [10,] 0.21955709  0.036948559 0.17097565  0.20942343  0.353240541 0.35675558
#> [11,] 0.17332004  0.053722020 0.12723476  0.18043010  0.187525473 0.18339501
#> [12,] 0.08940000  0.057000000 0.11270000  0.08730000  0.102000000 0.19640000
#> [13,] 0.44488701  0.069517367 0.34205334  0.36164376  0.356423371 0.32278518
#> [14,] 0.35658955  0.023519393 0.22905639  0.06779096  0.200672499 0.17539717
#> [15,] 0.27014576 -0.004726517 0.09613847 -0.04122038 -0.006055785 0.08074638
#> [16,] 0.23380000  0.017300000 0.23820000  0.33180000  0.382100000 0.15750000
#> [17,] 0.63184276  0.025899213 0.58648660  0.46137228  0.404128607 0.30689646
#> [18,] 0.03150887  0.031291356 0.64820720  0.77231453  0.615276556 0.33535123
#> [19,] 0.14016943  0.044137326 0.21934058  0.19258720  0.145767564 0.10882378
#> [20,] 0.22212280 -0.012467452 0.24503160  0.27031950  0.225498818 0.16233413
#> [21,] 0.25518927  0.020613408 0.14730497  0.12202728  0.114993719 0.10894241
#> [22,] 0.18289375  0.009776389 0.10857890  0.10288159  0.159683983 0.16707972
#> [23,] 0.91522774  0.019877482 0.78040263  0.69473424  0.594507992 0.48004082
#> [24,] 0.08492423  0.023076020 0.02540375  0.15497440  0.166611971 0.18658812
#> [25,] 0.22874483  0.035617920 0.27155522  0.22395164  0.229247045 0.25437089
#>              V44        V45         V46          V47         V48           V49
#>  [1,] 0.22290000 0.22030000  0.22650000  0.176600000  0.10970000  0.0558000000
#>  [2,] 0.19646768 0.10704515  0.14717581  0.155045344  0.09419978  0.0472686149
#>  [3,] 0.29498288 0.23096677  0.17112013  0.088375909  0.02780420  0.0349294174
#>  [4,] 0.48957008 0.50555474  0.41952907  0.303378141  0.20037142  0.1093964248
#>  [5,] 0.18180782 0.13562618  0.11285236  0.109235818  0.07880311  0.0506589164
#>  [6,] 0.26824291 0.25884247  0.22297131  0.177221928  0.11495718  0.0528411796
#>  [7,] 0.11802330 0.08228971  0.04705722  0.063994333  0.07322625  0.0453076141
#>  [8,] 0.20301111 0.21421154  0.22785627  0.193460473  0.13193442  0.0651028035
#>  [9,] 0.39412074 0.41728734  0.27408140  0.170690717  0.14484561  0.0848665930
#> [10,] 0.22999588 0.11519418  0.14136682  0.089639078  0.11510558  0.0788200758
#> [11,] 0.13619940 0.09489624  0.09010899  0.128950432  0.12824463  0.0684649588
#> [12,] 0.22560000 0.18140000  0.20120000  0.168800000  0.10370000  0.0501000000
#> [13,] 0.22723034 0.19175187  0.15767672  0.104108245  0.09678832  0.0509636859
#> [14,] 0.15746092 0.11487825  0.07208833  0.105138302  0.08915162  0.0385133349
#> [15,] 0.09507168 0.03620757 -0.01050373 -0.012247328 -0.01674032 -0.0009474142
#> [16,] 0.22280000 0.15820000  0.14330000  0.163400000  0.11330000  0.0567000000
#> [17,] 0.20946863 0.16703589  0.09590193  0.089881520  0.08254442  0.0377100531
#> [18,] 0.17321946 0.21245636  0.34372158  0.234936466  0.04841353  0.0032788372
#> [19,] 0.14389169 0.13767972  0.09182138  0.110022278  0.08711632  0.0469948753
#> [20,] 0.14775909 0.13461718  0.12251854  0.043472013  0.03008303  0.0279444214
#> [21,] 0.06667748 0.03669507  0.04181709  0.049094386  0.04042085  0.0215546064
#> [22,] 0.11957338 0.08085222  0.08772742  0.060865407  0.03434154  0.0079754612
#> [23,] 0.42177144 0.23119097  0.18003789  0.177074389  0.12346827  0.0606604341
#> [24,] 0.16940572 0.07845359  0.04949052  0.005321344  0.02357611  0.0075747316
#> [25,] 0.23654704 0.14050786  0.08501073  0.053316194  0.04505479  0.0121097647
#>                 V5          V50          V51           V52         V53
#>  [1,]  0.095800000  0.014200000  0.028100000  0.0165000000 0.005600000
#>  [2,]  0.169513576  0.026056840  0.025375263  0.0210540380 0.024096694
#>  [3,]  0.112255022 -0.002275461 -0.010620523 -0.0017327514 0.012848293
#>  [4,]  0.067112396  0.036919522  0.027025615  0.0225937995 0.011248626
#>  [5,]  0.097636072  0.023221782  0.016004714  0.0112196151 0.014055879
#>  [6,]  0.076159325  0.020396264  0.019899817  0.0116783906 0.008711769
#>  [7,]  0.069113441  0.024283184  0.021558753  0.0205965494 0.013917828
#>  [8,]  0.074461194  0.018766655  0.014533597  0.0150678046 0.008090586
#>  [9,]  0.093147789  0.026156974  0.024890919  0.0130288773 0.012472917
#> [10,]  0.120215120  0.014560122  0.028195326  0.0326744613 0.017644041
#> [11,]  0.071648359  0.017845036  0.014740484  0.0110598201 0.009357643
#> [12,]  0.052900000  0.013600000  0.013000000  0.0120000000 0.003900000
#> [13,]  0.129444086  0.019739084  0.017970667  0.0124237716 0.012014494
#> [14,]  0.006469145  0.014400772  0.002116902  0.0034958122 0.002949713
#> [15,]  0.027062059  0.005973502  0.031210850  0.0421507860 0.015880801
#> [16,]  0.047600000  0.013300000  0.017000000  0.0035000000 0.005200000
#> [17,]  0.017672833  0.016670243  0.016967968  0.0061526082 0.012023903
#> [18,]  0.070750053  0.048811634  0.044971975  0.0366922016 0.015775965
#> [19,]  0.080559718  0.017716880  0.012446088  0.0108081724 0.009974198
#> [20,]  0.007299237  0.008452148  0.001805831  0.0041373002 0.004592524
#> [21,]  0.050758129  0.014078628  0.002114180  0.0044600858 0.006927427
#> [22,]  0.041546789  0.022486175  0.005078054  0.0073914332 0.009503908
#> [23,]  0.046592039  0.021058689  0.014552275  0.0125123899 0.014267168
#> [24,] -0.018331248  0.011806008  0.009346594  0.0002437258 0.007417206
#> [25,]  0.064838283  0.009928168  0.010414409  0.0063171091 0.010232629
#>                V54         V55         V56          V57           V58
#>  [1,]  0.001000000 0.002700000 0.006200000 0.0024000000  0.0063000000
#>  [2,]  0.018944556 0.004199382 0.009261337 0.0111735579  0.0091510574
#>  [3,]  0.014636554 0.007285525 0.018368747 0.0130948536  0.0117019821
#>  [4,]  0.012078447 0.009503561 0.009667486 0.0094813498  0.0071696263
#>  [5,]  0.011836846 0.006715844 0.011841806 0.0070781546  0.0087138847
#>  [6,]  0.007472549 0.002441751 0.002750680 0.0047536273  0.0016446712
#>  [7,]  0.012342199 0.012844135 0.012217230 0.0033622931  0.0124879251
#>  [8,]  0.003650701 0.003875494 0.006996299 0.0040119655  0.0009717931
#>  [9,]  0.011947641 0.014245518 0.009488649 0.0082385862  0.0074115800
#> [10,]  0.032006284 0.005257876 0.002733009 0.0050841085  0.0180277458
#> [11,]  0.011764569 0.005515909 0.007443816 0.0051673010  0.0061854048
#> [12,]  0.005300000 0.006200000 0.004600000 0.0045000000  0.0022000000
#> [13,]  0.014753665 0.005448377 0.007277493 0.0042904910  0.0069549918
#> [14,]  0.003066709 0.000494182 0.003962422 0.0041750963  0.0016922511
#> [15,] -0.004064699 0.009262167 0.010078257 0.0083992675  0.0184247802
#> [16,]  0.008300000 0.007800000 0.007500000 0.0105000000  0.0160000000
#> [17,]  0.006841934 0.012941437 0.009323703 0.0039050917  0.0043789401
#> [18,]  0.014693364 0.009431907 0.002110915 0.0080386832 -0.0015627620
#> [19,]  0.007603232 0.005316123 0.009161559 0.0082313097  0.0046106541
#> [20,]  0.007260996 0.002207187 0.001655118 0.0014866417  0.0033165116
#> [21,]  0.006223906 0.006784402 0.003751726 0.0053396637  0.0052912770
#> [22,]  0.007413316 0.006618222 0.005519118 0.0080465301  0.0070047237
#> [23,]  0.003487650 0.004240456 0.003292459 0.0005847887  0.0152861463
#> [24,]  0.013455631 0.017555613 0.012540613 0.0094412292  0.0073742385
#> [25,]  0.011910999 0.007551966 0.009191228 0.0115139450  0.0048389737
#>               V59         V6          V60         V7          V8           V9
#>  [1,] 0.001700000 0.08300000  0.002800000 0.08790000 0.122000000  0.197700000
#>  [2,] 0.010918115 0.10283673  0.007792581 0.10851567 0.111722837  0.021166545
#>  [3,] 0.012213779 0.13926470  0.006415618 0.13310587 0.006281626  0.017384404
#>  [4,] 0.007580543 0.08953128  0.004629273 0.10848522 0.132015451  0.249065290
#>  [5,] 0.007410262 0.13183441  0.005202861 0.14642067 0.121470738  0.175109165
#>  [6,] 0.002763543 0.07899672  0.003663164 0.07328756 0.112774755  0.177196571
#>  [7,] 0.008263900 0.06027193  0.009448460 0.06980692 0.044378016  0.137956562
#>  [8,] 0.001659540 0.13713924  0.008388322 0.16514068 0.192680849  0.264928676
#>  [9,] 0.008538412 0.10313155  0.005564392 0.11396315 0.141530872  0.219868922
#> [10,] 0.009121891 0.22102254 -0.001603477 0.13164998 0.110126335  0.179483317
#> [11,] 0.007929999 0.10965261  0.007020047 0.11736814 0.132953165  0.167245551
#> [12,] 0.000500000 0.10910000  0.003100000 0.17090000 0.168400000  0.186500000
#> [13,] 0.009908709 0.07113631  0.003770782 0.06003775 0.153839759  0.349168823
#> [14,] 0.003016951 0.02950684  0.003413234 0.08179535 0.074244643  0.076685224
#> [15,] 0.009772514 0.15749736  0.003709280 0.13244086 0.227946131  0.150598548
#> [16,] 0.009500000 0.08160000  0.001100000 0.09930000 0.031500000  0.073600000
#> [17,] 0.005468043 0.06093159  0.004151094 0.05386826 0.052232341 -0.007933029
#> [18,] 0.004027604 0.11074609  0.003813886 0.16971162 0.130697258  0.115526291
#> [19,] 0.006305091 0.10239615  0.005399600 0.13653494 0.124934292  0.142070146
#> [20,] 0.001545891 0.00958960  0.002935806 0.05003652 0.080462761  0.086994224
#> [21,] 0.002216636 0.06224629  0.002728506 0.09143892 0.093048839  0.137041291
#> [22,] 0.004285201 0.05221407  0.003274169 0.06355448 0.031015193  0.054752610
#> [23,] 0.010612448 0.09082316  0.007068690 0.09976861 0.123452119  0.166237715
#> [24,] 0.004987498 0.06606760  0.006275943 0.13447819 0.177902666  0.210473557
#> [25,] 0.007112510 0.08239802  0.007458540 0.10512794 0.093108873  0.072648529
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768