Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.031387983 0.23437439 0.38269292 0.50256030 0.58722549 0.62310112
#>  [2,] 0.004907419 0.16633741 0.21299337 0.30757714 0.42486909 0.53411850
#>  [3,] 0.022778525 0.36810097 0.31604220 0.33445444 0.35016924 0.28766938
#>  [4,] 0.067393480 0.12551365 0.23929736 0.19395399 0.26558537 0.32114257
#>  [5,] 0.103797861 0.13869564 0.25449573 0.26703353 0.28241125 0.30092905
#>  [6,] 0.024102395 0.10198171 0.19349053 0.25787784 0.17465319 0.16160517
#>  [7,] 0.034849011 0.12877460 0.16850130 0.20736038 0.23576898 0.32787809
#>  [8,] 0.025454896 0.30033941 0.36358589 0.41080023 0.39152344 0.31381897
#>  [9,] 0.037873284 0.27155180 0.32714691 0.30162498 0.25113694 0.29408065
#> [10,] 0.040505179 0.22407145 0.29228201 0.32118317 0.27654500 0.19711288
#> [11,] 0.069291688 0.53273266 0.38813610 0.13712001 0.20385683 0.30933626
#> [12,] 0.026342599 0.22364885 0.25489108 0.29643709 0.32646532 0.37763458
#> [13,] 0.056780597 0.31497619 0.37131810 0.24621360 0.51483510 0.91793655
#> [14,] 0.023124578 0.09461111 0.14191655 0.08031309 0.14447292 0.18988433
#> [15,] 0.032855542 0.06906896 0.12429919 0.19861764 0.22653780 0.20894392
#> [16,] 0.028494166 0.26481617 0.27509933 0.22161520 0.29877368 0.34512982
#> [17,] 0.028965734 0.05641647 0.04356231 0.09723594 0.18799858 0.33590061
#> [18,] 0.008431805 0.11486845 0.06301724 0.08589478 0.07856813 0.07199736
#> [19,] 0.016381632 0.17169821 0.13865039 0.07674054 0.07007536 0.11671771
#> [20,] 0.009897355 0.12591514 0.13114374 0.09929896 0.12330573 0.12055309
#> [21,] 0.007014008 0.02945360 0.05022948 0.08262646 0.12951767 0.11092194
#> [22,] 0.043799565 0.26028373 0.26631120 0.32638546 0.28250698 0.26260185
#> [23,] 0.014213737 0.02171846 0.14745021 0.34340504 0.27357274 0.20770262
#> [24,] 0.016539206 0.09649298 0.09978922 0.11831780 0.10655170 0.11561861
#> [25,] 0.022414590 0.13908710 0.12114173 0.09645139 0.18494422 0.33949770
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.69287562 0.79645354 0.9314410 0.9681593 0.9109505 0.039901809 0.8395378
#>  [2,] 0.53375310 0.48915572 0.6480843 0.7497302 0.8032276 0.031431567 0.8749640
#>  [3,] 0.26364934 0.24713887 0.1967875 0.1427874 0.2802183 0.022021632 0.3647131
#>  [4,] 0.42593174 0.50619100 0.5016570 0.3363413 0.4271564 0.119868020 0.5536346
#>  [5,] 0.38668688 0.31347336 0.3999269 0.2078444 0.3178020 0.083851153 0.3371068
#>  [6,] 0.24045440 0.34150464 0.4423856 0.3884425 0.4350822 0.057016814 0.4381000
#>  [7,] 0.39946908 0.49262219 0.5743249 0.6905525 0.8057956 0.046500180 0.8800080
#>  [8,] 0.24988205 0.27309558 0.3118166 0.3529894 0.5106379 0.041904783 0.6664341
#>  [9,] 0.23160474 0.21341175 0.1923725 0.1773224 0.2837887 0.024715336 0.2932327
#> [10,] 0.16321782 0.16254182 0.1754408 0.2126265 0.2560833 0.036032908 0.2926682
#> [11,] 0.43315807 0.44512069 0.1867115 0.1528589 0.3112074 0.089914513 0.8813989
#> [12,] 0.44217285 0.52579674 0.6169234 0.6741475 0.7115620 0.029362992 0.7807126
#> [13,] 0.94514199 0.89048495 0.7667628 0.8404759 0.9197213 0.064466602 0.8600846
#> [14,] 0.21225314 0.18192427 0.1443303 0.1623731 0.1541603 0.017346216 0.2249055
#> [15,] 0.25434834 0.37723416 0.5816982 0.4952564 0.3630858 0.043257798 0.3748631
#> [16,] 0.45963866 0.57800039 0.6684382 0.6650723 0.5561844 0.050353679 0.4953799
#> [17,] 0.45918465 0.60421944 0.7007147 0.7507411 0.7612523 0.025976580 0.8102165
#> [18,] 0.05447309 0.12374281 0.2449458 0.2922383 0.3579849 0.019822244 0.4839461
#> [19,] 0.12835201 0.16867514 0.2438250 0.2521290 0.3441556 0.000508184 0.5355634
#> [20,] 0.12501636 0.14300616 0.1894214 0.2103106 0.3398526 0.010916488 0.3940939
#> [21,] 0.04948836 0.07965229 0.1090669 0.1482838 0.2117372 0.005115431 0.2057193
#> [22,] 0.36276736 0.41782966 0.4672343 0.5707400 0.7120240 0.068617053 0.8445888
#> [23,] 0.07737416 0.08591989 0.1831398 0.4121738 0.2829234 0.047999949 0.5679838
#> [24,] 0.09219152 0.17382718 0.1703148 0.3217449 0.3442296 0.028085057 0.2802030
#> [25,] 0.57145837 0.78436367 0.8567304 0.8340824 0.8628162 0.014604446 0.8537812
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.7167758 0.5989903 0.4744108 0.3766550 0.3039455 0.2027615 0.1823082
#>  [2,] 1.0123796 1.0387996 0.9036482 0.9031475 0.9792277 0.9358798 0.7772283
#>  [3,] 0.3741067 0.3351244 0.4407617 0.5022548 0.6198971 0.7202802 0.7176717
#>  [4,] 0.7069814 0.7541515 0.5958902 0.6353426 0.6418880 0.6815409 0.7318405
#>  [5,] 0.3144131 0.2834153 0.3609178 0.2494841 0.4082554 0.4566021 0.4737912
#>  [6,] 0.6299629 0.6016523 0.6235676 0.6081855 0.4462353 0.5757012 0.7814733
#>  [7,] 0.9356380 1.0067708 0.9666366 0.8840115 0.7464735 0.6140148 0.5732080
#>  [8,] 0.7744896 0.8096767 0.8062845 0.8954808 0.9262343 0.9428101 0.9769644
#>  [9,] 0.3327593 0.3405637 0.4100425 0.4322010 0.4360979 0.6323779 0.8848826
#> [10,] 0.4041718 0.5431857 0.6823198 0.7647712 0.8282927 0.8561809 0.8878799
#> [11,] 0.9540153 0.3932707 0.1412744 0.1869737 0.1605494 0.3598735 0.2679831
#> [12,] 0.8415707 0.8824792 0.9335456 0.9893636 0.9925443 0.9712339 0.9337379
#> [13,] 0.8516491 0.8962777 0.9645688 0.8972826 0.7521808 0.7500242 0.5904381
#> [14,] 0.3130420 0.4793047 0.7311482 0.8164924 0.8082407 0.8203902 0.7809660
#> [15,] 0.4029402 0.5181027 0.4937538 0.5590942 0.7023796 0.7389937 0.6983677
#> [16,] 0.5191535 0.4698573 0.5451688 0.7411078 0.6956037 0.6348444 0.7281976
#> [17,] 0.9915638 0.9665054 0.9405953 0.9692943 0.9413766 0.9362438 0.8131807
#> [18,] 0.6082983 0.7869501 0.8801204 0.8268660 0.7340400 0.7972547 0.9339081
#> [19,] 0.6484337 0.8318652 0.9815998 0.9097264 0.8697551 0.9426376 0.9483691
#> [20,] 0.3554868 0.3915241 0.5917434 0.7021682 0.7960606 0.8473308 0.8255076
#> [21,] 0.2152650 0.3489018 0.4223712 0.5634159 0.6139750 0.6525965 0.6301066
#> [22,] 0.9032408 0.8236305 0.7313806 0.7019112 0.7377709 0.7786823 0.7155426
#> [23,] 0.7157001 0.7317068 0.7688710 0.8165689 0.9980498 0.9130552 0.7275202
#> [24,] 0.2344865 0.1788785 0.2259227 0.2548687 0.1138619 0.2552781 0.3779491
#> [25,] 0.8976596 0.9124734 0.8894652 0.9217000 0.9220194 0.9387193 0.9343488
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.1711327 0.1222893 0.051423502 0.1673691 0.2552478 0.3403937 0.33436416
#>  [2,] 0.6369042 0.5364225 0.030395511 0.3707587 0.2141289 0.1180205 0.10250583
#>  [3,] 0.8213801 0.8397689 0.038570240 0.8725131 0.9003248 0.7373760 0.55883618
#>  [4,] 0.8130839 0.8572906 0.130417911 0.9170935 0.9537756 0.8198065 0.48415319
#>  [5,] 0.7358626 0.7611134 0.058120020 0.6532776 0.7502044 0.6723425 0.49368683
#>  [6,] 0.7582272 0.6885656 0.065735568 0.8097690 0.5876617 0.4692890 0.31403197
#>  [7,] 0.4452730 0.3811096 0.061506775 0.3475982 0.2460627 0.2348948 0.36447005
#>  [8,] 0.9488813 0.8169862 0.040895082 0.6076025 0.4572461 0.4239609 0.34980832
#>  [9,] 0.9956220 1.0298680 0.002716203 0.8562607 0.5432467 0.4889752 0.52169714
#> [10,] 0.9335693 0.8900411 0.050788666 0.7335898 0.7017360 0.5381990 0.40536023
#> [11,] 0.4925476 0.4710912 0.087177006 0.6944240 0.2912965 0.3511544 0.58541447
#> [12,] 0.8636762 0.7310404 0.032453440 0.5742629 0.4220096 0.3350574 0.23620866
#> [13,] 0.4700195 0.5502200 0.041874112 0.3823034 0.3154394 0.1714446 0.06636565
#> [14,] 0.8509207 0.9340661 0.025491663 0.8079674 0.6619911 0.5531543 0.62385966
#> [15,] 0.6884315 0.7162734 0.043129738 0.6840168 0.6792415 0.6833289 0.70589178
#> [16,] 0.7880403 0.6175296 0.064458578 0.5347611 0.5090651 0.3863395 0.25952298
#> [17,] 0.5724158 0.3246078 0.033489424 0.4042650 0.4500491 0.3228085 0.27580248
#> [18,] 0.9575799 0.9639222 0.017834721 0.8586210 0.7795017 0.7316889 0.68415474
#> [19,] 0.9001405 0.8237193 0.001444124 0.7176081 0.6497865 0.5654535 0.52000103
#> [20,] 0.7016882 0.5455801 0.015362345 0.3981556 0.2665637 0.2466615 0.24698841
#> [21,] 0.7770464 0.9823104 0.012310298 0.9486847 0.8013961 0.7689293 0.78482170
#> [22,] 0.5627007 0.3143686 0.048457706 0.1032812 0.2775233 0.6806877 0.73916458
#> [23,] 0.5655945 0.5044386 0.056931240 0.5029631 0.5490969 0.5590540 0.64589251
#> [24,] 0.5769711 0.7137717 0.025300176 0.6846831 0.5626692 0.3678529 0.54750153
#> [25,] 0.7615659 0.5343444 0.032606680 0.4210446 0.4332174 0.4001029 0.33530453
#>                V34        V35        V36       V37       V38        V39
#>  [1,]  0.360965124 0.37487239 0.23121939 0.1974623 0.3040069 0.27812628
#>  [2,] -0.006175456 0.01144928 0.19711765 0.2231415 0.3345105 0.42935413
#>  [3,]  0.416756115 0.25986309 0.25682641 0.2960631 0.3026124 0.41197657
#>  [4,]  0.454138124 0.53602081 0.21312875 0.3293303 0.3426462 0.25882503
#>  [5,]  0.598133428 0.65470220 0.39406853 0.1079684 0.3160663 0.29793938
#>  [6,]  0.333923189 0.22389074 0.09795135 0.2752241 0.3749182 0.24012226
#>  [7,]  0.333669992 0.32705532 0.39938865 0.4313885 0.3827697 0.39149601
#>  [8,]  0.245834250 0.12928487 0.19597196 0.2206792 0.2454792 0.29650398
#>  [9,]  0.369957396 0.14495712 0.11654495 0.2087467 0.1485994 0.03258934
#> [10,]  0.281529630 0.15727384 0.10129995 0.1856203 0.1314878 0.17749457
#> [11,]  0.754442626 0.84721350 0.87512873 0.8216964 0.9058866 0.73623236
#> [12,]  0.136794381 0.12390224 0.13616587 0.2154563 0.2590497 0.22661321
#> [13,]  0.159127829 0.21685339 0.19051834 0.1576641 0.1671729 0.29798484
#> [14,]  0.707606332 0.63506992 0.46438115 0.2276045 0.1412860 0.18399165
#> [15,]  0.794543134 0.87313416 0.91579420 0.8450869 0.7214839 0.51029041
#> [16,]  0.128989097 0.32336092 0.48965428 0.4380398 0.2566138 0.19528248
#> [17,]  0.261042258 0.14405894 0.04976637 0.1115982 0.1106063 0.11168456
#> [18,]  0.693809823 0.68633788 0.61207240 0.5341321 0.4382295 0.32581313
#> [19,]  0.555535371 0.56210566 0.48772197 0.4615389 0.3429522 0.17395454
#> [20,]  0.253016772 0.23533853 0.27374357 0.3143430 0.2678821 0.31704644
#> [21,]  0.616017418 0.49548651 0.41848349 0.3873028 0.3406175 0.32618688
#> [22,]  0.775145435 0.75114511 0.76041245 0.6982650 0.2659173 0.36397546
#> [23,]  0.568277215 0.58243395 0.59733523 0.4958869 0.2769323 0.05278762
#> [24,]  0.543375037 0.41557433 0.49692050 0.6077230 0.7139646 0.76746009
#> [25,]  0.303306100 0.25247933 0.14213877 0.2577704 0.2117050 0.13130986
#>                V4        V40        V41        V42        V43         V44
#>  [1,] 0.073735561 0.11582780 0.07741228 0.16149220 0.16135606 0.058278879
#>  [2,] 0.037988389 0.29835683 0.19987977 0.18754845 0.10351954 0.156251419
#>  [3,] 0.036047231 0.53873642 0.54867024 0.47146289 0.48963877 0.482806890
#>  [4,] 0.143124016 0.08655034 0.56086784 0.71760124 0.44720225 0.329497574
#>  [5,] 0.063562264 0.11845473 0.50502860 0.80044675 0.75121998 0.545790867
#>  [6,] 0.065748788 0.33902725 0.33544907 0.18434214 0.30817488 0.311954108
#>  [7,] 0.079792055 0.43886963 0.40253928 0.44048079 0.34151646 0.206979912
#>  [8,] 0.030773811 0.25074879 0.13188973 0.21417793 0.29357211 0.332927115
#>  [9,] 0.068089591 0.18745598 0.39540090 0.49543399 0.45796848 0.232120736
#> [10,] 0.057892785 0.19060039 0.23374549 0.25271767 0.17805858 0.088481206
#> [11,] 0.090104444 0.49995564 0.37609555 0.27495776 0.19721888 0.239900244
#> [12,] 0.046170955 0.23459123 0.24916922 0.23344007 0.20120938 0.178497130
#> [13,] 0.049147184 0.31250335 0.18712994 0.15843087 0.27958348 0.254839607
#> [14,] 0.036671552 0.23120826 0.19684256 0.23833596 0.29450658 0.180645349
#> [15,] 0.024405384 0.35975278 0.28390565 0.26257935 0.20635652 0.162056907
#> [16,] 0.062065099 0.25709282 0.26242938 0.23210549 0.16100765 0.119832801
#> [17,] 0.048439086 0.18734093 0.12327391 0.06029483 0.03647649 0.118381619
#> [18,] 0.036879106 0.23451022 0.16883710 0.09050747 0.05753162 0.057301696
#> [19,] 0.028838178 0.10825288 0.11336541 0.02411062 0.05482878 0.004224784
#> [20,] 0.006687523 0.27459100 0.20402268 0.18104396 0.16136258 0.169618804
#> [21,] 0.031274424 0.28821825 0.14992445 0.03439290 0.08903142 0.164579645
#> [22,] 0.077740440 0.52165554 0.49086847 0.42583268 0.36669707 0.402220717
#> [23,] 0.048435338 0.31864715 0.39645893 0.55737852 0.47655688 0.282696251
#> [24,] 0.018412445 0.65347679 0.43238315 0.37361666 0.27562094 0.169400706
#> [25,] 0.025686020 0.13394714 0.08280427 0.05859109 0.09360098 0.106684679
#>               V45        V46        V47        V48         V49         V5
#>  [1,]  0.08640775 0.05859624 0.03808998 0.04253850 0.026088861 0.10099189
#>  [2,]  0.13916057 0.11164947 0.12274924 0.10620356 0.077988798 0.14103833
#>  [3,]  0.62789941 0.51702582 0.37232494 0.26319965 0.146397474 0.06288877
#>  [4,]  0.59411202 0.57823667 0.26968229 0.12276803 0.087584106 0.15721649
#>  [5,]  0.58546443 0.73537363 0.55759137 0.33718905 0.162248960 0.06198882
#>  [6,]  0.17638394 0.12660096 0.07046812 0.08944037 0.081019093 0.03212456
#>  [7,]  0.11470081 0.08476164 0.09054580 0.06461184 0.036034500 0.11030516
#>  [8,]  0.30159853 0.18845505 0.12408630 0.10475862 0.057143820 0.05750490
#>  [9,]  0.24788522 0.25987665 0.21817903 0.18810813 0.124492361 0.08282954
#> [10,]  0.09895129 0.15191124 0.17429831 0.13705069 0.059256625 0.07275016
#> [11,]  0.40191086 0.29809810 0.16564859 0.17715775 0.090198131 0.05592818
#> [12,]  0.13713104 0.09600565 0.10284847 0.07980763 0.047110876 0.10067966
#> [13,]  0.12976135 0.12033129 0.14079639 0.12013535 0.093166380 0.06320942
#> [14,]  0.05681422 0.05188357 0.04700548 0.03852522 0.011812229 0.04002887
#> [15,]  0.14608279 0.16500546 0.11350276 0.04697617 0.014762451 0.03194085
#> [16,]  0.12321251 0.10553746 0.08850178 0.07397849 0.043014925 0.07218617
#> [17,]  0.09433038 0.04292024 0.08422142 0.05982074 0.026871741 0.06893016
#> [18,]  0.01030361 0.01696756 0.02074189 0.03596311 0.011749708 0.05942357
#> [19,] -0.01802919 0.01159813 0.02346023 0.02715901 0.012998681 0.05427974
#> [20,]  0.11377548 0.06357108 0.01956733 0.02157900 0.018663245 0.02484858
#> [21,]  0.19008101 0.09307283 0.03673344 0.02509663 0.009373368 0.02592163
#> [22,]  0.18744691 0.06893483 0.06364700 0.03118961 0.041081628 0.08662536
#> [23,]  0.19445928 0.15969442 0.09022268 0.07416175 0.023018144 0.12553577
#> [24,]  0.16491861 0.14372563 0.14460646 0.10554340 0.054555559 0.04274713
#> [25,]  0.11502842 0.07204983 0.06209663 0.04739721 0.024645711 0.06776194
#>               V50         V51         V52           V53           V54
#>  [1,] 0.015901766 0.020413713 0.013089738  1.541751e-02  0.0178064876
#>  [2,] 0.029292476 0.023945852 0.016888959  1.227394e-02  0.0268822057
#>  [3,] 0.045243916 0.011558851 0.011646985  1.486309e-02  0.0085320620
#>  [4,] 0.037071079 0.031956418 0.020378008  2.103299e-02  0.0137942976
#>  [5,] 0.048097807 0.102173420 0.072262289  3.206446e-02  0.0313733564
#>  [6,] 0.020328824 0.021097560 0.022646119  1.258601e-02  0.0159425697
#>  [7,] 0.020250541 0.020046803 0.017256459  1.971813e-02  0.0186529329
#>  [8,] 0.011348296 0.022840307 0.013190283  5.453495e-03  0.0073916107
#>  [9,] 0.022276962 0.014996079 0.009469969 -2.136565e-05 -0.0004003941
#> [10,] 0.017417347 0.016813673 0.016081419  8.724448e-03  0.0093756894
#> [11,] 0.030362578 0.030471514 0.020524727  1.356879e-02  0.0240003389
#> [12,] 0.021850692 0.016846118 0.014301177  1.682238e-02  0.0121235900
#> [13,] 0.028082019 0.032057376 0.009366982  1.583846e-02  0.0156287701
#> [14,] 0.009382188 0.012223849 0.008218618  7.998405e-03  0.0067093311
#> [15,] 0.023181205 0.014588647 0.011895538  1.118064e-02  0.0085041272
#> [16,] 0.016206993 0.011329921 0.014243278  7.456919e-03  0.0119153591
#> [17,] 0.018034365 0.007164782 0.012661385  1.077377e-02  0.0044482033
#> [18,] 0.010414873 0.004487307 0.006328443  4.259183e-03  0.0016736559
#> [19,] 0.008786709 0.003993615 0.003687793  8.559185e-03  0.0080195238
#> [20,] 0.005868232 0.005553522 0.012267897  8.104278e-03  0.0070086365
#> [21,] 0.008382671 0.007461994 0.004520726  1.072178e-02  0.0078797093
#> [22,] 0.021147321 0.005627313 0.009482319  1.402924e-02  0.0185824843
#> [23,] 0.009643368 0.005808695 0.001120244  8.810653e-03  0.0070535823
#> [24,] 0.023115749 0.012218145 0.016080464  2.137153e-02  0.0156276243
#> [25,] 0.008428198 0.016856606 0.009671048  7.519369e-03  0.0089879941
#>               V55         V56           V57          V58         V59         V6
#>  [1,] 0.018021427 0.010287811  0.0061970143 0.0134405513 0.008759914 0.10258820
#>  [2,] 0.016846296 0.005210524  0.0084048195 0.0058782432 0.009220806 0.18955070
#>  [3,] 0.004984966 0.004873086  0.0067938160 0.0066078717 0.005536930 0.12703397
#>  [4,] 0.012578819 0.013644195  0.0113437866 0.0127191443 0.014015070 0.11187355
#>  [5,] 0.025468902 0.008453321  0.0180377943 0.0214333350 0.022637775 0.13281111
#>  [6,] 0.014092759 0.013235222  0.0083506274 0.0088146374 0.007222992 0.08191429
#>  [7,] 0.001956009 0.010982170  0.0092508394 0.0088269322 0.010802914 0.10955495
#>  [8,] 0.007945042 0.004590837  0.0046007164 0.0041968287 0.004351152 0.09522846
#>  [9,] 0.004430358 0.010279913 -0.0007195881 0.0105379676 0.012403384 0.04743276
#> [10,] 0.006734644 0.004207303  0.0069926039 0.0076687985 0.004511646 0.12381827
#> [11,] 0.017380357 0.020868187  0.0160300135 0.0167185897 0.017551823 0.12880015
#> [12,] 0.006242605 0.011304800  0.0087188506 0.0087336082 0.006364013 0.13794847
#> [13,] 0.027620921 0.010217562  0.0085478417 0.0226600693 0.011823740 0.14169256
#> [14,] 0.005473363 0.006231772  0.0102435587 0.0083183544 0.007871909 0.06985817
#> [15,] 0.014572072 0.006331895  0.0096116166 0.0093380994 0.010190764 0.08371711
#> [16,] 0.010629412 0.007328416  0.0095729342 0.0060806406 0.007442547 0.10444170
#> [17,] 0.004688119 0.014398596  0.0110075503 0.0012147357 0.005648687 0.07599852
#> [18,] 0.001831551 0.002486374  0.0059514424 0.0052436258 0.002856080 0.07636760
#> [19,] 0.006977254 0.003493988  0.0038013868 0.0026017299 0.002288166 0.09382944
#> [20,] 0.005822348 0.002510081  0.0041929652 0.0088277486 0.004785461 0.03540951
#> [21,] 0.004245382 0.007061627  0.0047965844 0.0008248479 0.006100577 0.01182475
#> [22,] 0.006611407 0.017665926  0.0180875823 0.0106783309 0.005999785 0.21954480
#> [23,] 0.010571667 0.008719206  0.0095649991 0.0061338875 0.004070966 0.18021193
#> [24,] 0.017766078 0.007500990  0.0039410980 0.0077199347 0.008217995 0.06971031
#> [25,] 0.006014861 0.007187484  0.0054674361 0.0044330565 0.005845753 0.12285269
#>               V60         V7         V8         V9
#>  [1,] 0.007094990 0.12975454 0.13228106 0.12908413
#>  [2,] 0.005590512 0.19293541 0.18262464 0.20369174
#>  [3,] 0.003286645 0.12652315 0.12860514 0.28288637
#>  [4,] 0.019376373 0.08988494 0.11324075 0.17609635
#>  [5,] 0.010435548 0.09742216 0.17790441 0.12068226
#>  [6,] 0.009651564 0.11038832 0.05385424 0.12976238
#>  [7,] 0.009990449 0.09286866 0.13819248 0.10420571
#>  [8,] 0.004209828 0.10449732 0.15347327 0.23336108
#>  [9,] 0.004685393 0.06792794 0.13672477 0.22260872
#> [10,] 0.004398333 0.12047094 0.11014948 0.15145600
#> [11,] 0.012595356 0.22637254 0.40880825 0.59101041
#> [12,] 0.004972965 0.13557497 0.14656637 0.19958808
#> [13,] 0.005912923 0.24019954 0.26474364 0.25959655
#> [14,] 0.006558138 0.06913416 0.08934285 0.08188714
#> [15,] 0.007279893 0.12566725 0.10376092 0.05075545
#> [16,] 0.005306928 0.12739721 0.15542635 0.22734697
#> [17,] 0.004764290 0.11690434 0.11001046 0.10319030
#> [18,] 0.002919278 0.04539469 0.10068830 0.11432740
#> [19,] 0.002362934 0.10099450 0.09376214 0.16764329
#> [20,] 0.003793719 0.06065609 0.08834095 0.09337374
#> [21,] 0.002633704 0.04113319 0.03083779 0.01108940
#> [22,] 0.004819612 0.15272807 0.11644199 0.10080318
#> [23,] 0.014215464 0.14486418 0.14090328 0.02895304
#> [24,] 0.007979480 0.07634396 0.11259970 0.09207117
#> [25,] 0.006075432 0.11413966 0.06935949 0.13364333
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696