Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1          V10        V11          V12        V13        V14
#>  [1,] 0.035474811  0.246504642 0.27518658  0.284111157 0.27531353 0.22796399
#>  [2,] 0.017835226  0.306235261 0.26136664  0.318523956 0.28456687 0.18908118
#>  [3,] 0.031002876  0.303546795 0.34953488  0.321962793 0.23668427 0.13688088
#>  [4,] 0.035858313  0.219104940 0.22345867  0.180369197 0.19436420 0.29057771
#>  [5,] 0.080029725  0.661709349 0.49760423  0.249568787 0.15740578 0.43280242
#>  [6,] 0.088551265  0.524607850 0.36524263  0.109092959 0.14124508 0.25104321
#>  [7,] 0.076947719  0.247278742 0.36334495  0.445686668 0.54009515 0.55136761
#>  [8,] 0.016450456  0.246948633 0.28976519  0.335097064 0.29889325 0.22327670
#>  [9,] 0.030435445  0.251096072 0.29087560  0.300294135 0.36964606 0.47287059
#> [10,] 0.044389473  0.001827808 0.11656334  0.201956828 0.25871834 0.32480423
#> [11,] 0.042858337  0.040811108 0.22374689  0.239184760 0.18808447 0.35390816
#> [12,] 0.030294843  0.352964565 0.42395980  0.445136488 0.41288455 0.32608107
#> [13,] 0.017161636  0.160496895 0.08429198  0.017235532 0.09291801 0.23972394
#> [14,] 0.018867497  0.136837287 0.15000712  0.114955014 0.16181694 0.23417617
#> [15,] 0.008954821  0.121263455 0.10827663 -0.007901333 0.03221467 0.04865392
#> [16,] 0.016623827  0.090509835 0.08521116  0.060525089 0.07704341 0.10151601
#> [17,] 0.019940602  0.112197242 0.10219172  0.104295232 0.12651027 0.06389071
#> [18,] 0.012414688  0.127790436 0.10162265  0.113687650 0.08732430 0.10254006
#> [19,] 0.037198395  0.334690841 0.29520707  0.176333425 0.25305649 0.47920902
#> [20,] 0.020975105  0.095930584 0.08457637  0.068670565 0.17487088 0.31646590
#> [21,] 0.016524127  0.161378579 0.12437248  0.161266558 0.30793098 0.23446474
#> [22,] 0.030477527 -0.044862058 0.08362372  0.357501666 0.34071113 0.30754168
#> [23,] 0.021944773  0.128830095 0.12182977  0.213259703 0.20816631 0.15170134
#> [24,] 0.024782916  0.127855220 0.15217266  0.175051500 0.16512061 0.16496642
#> [25,] 0.005850630  0.112374136 0.12586180  0.071427471 0.05611372 0.11465260
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.22807077 0.22908552 0.24290299 0.1622057 0.1732495 0.041178616
#>  [2,] 0.09499440 0.20147534 0.23293431 0.2815721 0.2351351 0.017127649
#>  [3,] 0.12103143 0.13041418 0.17685957 0.2670859 0.3972728 0.055880017
#>  [4,] 0.41932286 0.49444613 0.60417889 0.5215087 0.5809691 0.045107742
#>  [5,] 0.44766973 0.45331304 0.32998284 0.3930747 0.5254713 0.070943780
#>  [6,] 0.39369079 0.39747435 0.25357095 0.2093067 0.3351795 0.097804209
#>  [7,] 0.66052756 0.77426697 0.93858752 0.9221204 0.8769587 0.120241831
#>  [8,] 0.16463345 0.20772601 0.24303987 0.3762367 0.6023331 0.028893603
#>  [9,] 0.52786572 0.58568819 0.62788189 0.7056531 0.7820384 0.037722023
#> [10,] 0.40096706 0.49934235 0.64837334 0.7202827 0.8538550 0.038083388
#> [11,] 0.42442355 0.50547699 0.41623418 0.3685048 0.4037621 0.087694347
#> [12,] 0.24289247 0.28200747 0.30844125 0.3344308 0.4429728 0.035521122
#> [13,] 0.40144577 0.60233656 0.71775068 0.7937591 0.8936764 0.015954617
#> [14,] 0.23604559 0.17407380 0.15594064 0.2060929 0.2287413 0.018981949
#> [15,] 0.12460753 0.06523489 0.09826287 0.2048468 0.3422132 0.016303790
#> [16,] 0.03771791 0.13326107 0.17051921 0.3896277 0.3468226 0.016034234
#> [17,] 0.01962430 0.03554034 0.07839297 0.2052348 0.3553955 0.005681703
#> [18,] 0.11298511 0.17167052 0.25429756 0.3994076 0.4406477 0.015626792
#> [19,] 0.61754289 0.73398720 0.78892622 0.7115146 0.5166484 0.061931944
#> [20,] 0.51158942 0.69854629 0.76665100 0.7697431 0.8271479 0.019272720
#> [21,] 0.16767171 0.30554679 0.31726075 0.3459019 0.5772231 0.035314331
#> [22,] 0.03360967 0.04807710 0.16980228 0.4413546 0.2843149 0.020325398
#> [23,] 0.18798153 0.23141934 0.21701199 0.2787283 0.2957209 0.037618086
#> [24,] 0.22614967 0.50822866 0.70972481 0.5989751 0.3244362 0.016335251
#> [25,] 0.16475741 0.15270408 0.12913295 0.1620351 0.2447772 0.000696752
#>             V20        V21       V22        V23        V24        V25       V26
#>  [1,] 0.2114576 0.29703459 0.4242845 0.54242885 0.61471778 0.65242173 0.6952769
#>  [2,] 0.1608909 0.07007036 0.1765322 0.40310339 0.56820118 0.66703802 0.7621117
#>  [3,] 0.4528617 0.56917006 0.6555187 0.72321414 0.80275361 0.82873771 0.9350189
#>  [4,] 0.7925375 0.97497721 0.9716447 0.73890393 0.48442735 0.32030453 0.4810479
#>  [5,] 0.9017625 0.99773222 0.3629557 0.09480127 0.01816919 0.22592393 0.4011580
#>  [6,] 0.8553760 0.98196242 0.4251148 0.09668345 0.20437218 0.14497511 0.3159395
#>  [7,] 0.7542385 0.75593644 0.5920690 0.36056577 0.27970796 0.24787442 0.2256168
#>  [8,] 0.7099547 0.80899676 0.8001119 0.82988260 0.93845457 0.93639775 0.9526826
#>  [9,] 0.8358009 0.87524917 0.9114381 0.95129737 0.95270245 0.88336162 0.8874185
#> [10,] 0.9297956 0.93103118 0.9244991 0.93203533 0.91403040 0.76205861 0.6196860
#> [11,] 0.2744769 0.22091639 0.1083268 0.27275251 0.54742948 0.42138010 0.4038717
#> [12,] 0.6185261 0.72510342 0.8011306 0.80305465 0.82857526 0.82518471 0.8924489
#> [13,] 0.9415951 0.91872192 0.9464197 0.86439223 0.55485879 0.31969558 0.3060764
#> [14,] 0.3423906 0.42435422 0.5220495 0.77061465 0.86318798 0.85002912 0.8543535
#> [15,] 0.3590057 0.46170504 0.8863849 0.97274988 0.88859165 0.68126693 0.6923756
#> [16,] 0.3119763 0.21828159 0.1868520 0.17946857 0.22354369 0.05091038 0.1854291
#> [17,] 0.3771822 0.34190654 0.4191760 0.57497923 0.70780128 0.78695067 0.7900682
#> [18,] 0.5164398 0.59257028 0.6540544 0.64905610 0.61296093 0.50498110 0.5128846
#> [19,] 0.4361741 0.56225134 0.6849156 0.76326288 0.79597287 0.73649819 0.6081196
#> [20,] 0.8563750 0.96574763 0.9594217 0.94117586 0.94232653 0.89695812 0.9295369
#> [21,] 0.6607725 0.63358077 0.5293569 0.49100434 0.43687530 0.57361551 0.7459034
#> [22,] 0.5393840 0.67759716 0.7050187 0.78827240 0.83531598 0.97771403 0.8343318
#> [23,] 0.1975151 0.24863700 0.1916002 0.24833289 0.26867028 0.21438648 0.3058355
#> [24,] 0.2849581 0.39824697 0.5240627 0.57640933 0.63932946 0.81361610 0.9483606
#> [25,] 0.3327556 0.35714277 0.4264916 0.63091024 0.76045948 0.83231860 0.8205435
#>             V27       V28       V29           V3       V30       V31       V32
#>  [1,] 0.7666656 0.8545310 0.8670167  0.051364458 0.8698522 0.8767544 0.8109699
#>  [2,] 0.8976027 0.9341131 0.9084725  0.025093331 0.9707861 0.9790006 0.8883513
#>  [3,] 0.9997072 0.9627950 0.8352627  0.039452194 0.6610685 0.5579622 0.4611596
#>  [4,] 0.5424210 0.2999710 0.2193185  0.029587860 0.4738489 0.2838999 0.4002040
#>  [5,] 0.2204179 0.5182915 0.6914811  0.033743173 0.8312139 0.4244363 0.3715569
#>  [6,] 0.2149456 0.3510719 0.4415704  0.087998391 0.6439395 0.2555858 0.3193143
#>  [7,] 0.2251052 0.2481946 0.1620672  0.124146185 0.2105769 0.1953015 0.2406480
#>  [8,] 0.9812164 0.9818105 0.8671823  0.034835821 0.6276725 0.3926353 0.2886952
#>  [9,] 0.8897881 0.8089634 0.6602116  0.037586165 0.5136899 0.3849505 0.2800527
#> [10,] 0.5336380 0.4260048 0.4496216  0.056781621 0.4209706 0.2941193 0.1763643
#> [11,] 0.4989018 0.4966076 0.4638529  0.117970894 0.4349698 0.3760135 0.3410474
#> [12,] 0.9154593 0.8034196 0.6394487  0.042042434 0.4904598 0.3794768 0.3951340
#> [13,] 0.3192702 0.4841560 0.6702121  0.002282089 0.6746699 0.4375980 0.3567540
#> [14,] 0.8236386 0.8608623 0.9175951  0.024821348 0.7878120 0.6371901 0.5372332
#> [15,] 0.5723069 0.6580909 0.7900568  0.041947705 0.7668233 0.5827707 0.6289937
#> [16,] 0.4154900 0.6491257 0.7941321  0.004461272 0.7242503 0.5063248 0.1751761
#> [17,] 0.7069737 0.7969131 0.9675100  0.014942355 0.9720218 0.9318356 0.8884935
#> [18,] 0.6002098 0.7539658 0.9030573  0.008300138 0.8976275 0.8082774 0.6609256
#> [19,] 0.6474795 0.6112642 0.3924923  0.079768157 0.4350922 0.5016007 0.3926857
#> [20,] 0.8869610 0.6971458 0.4520555  0.029155777 0.4200501 0.4240945 0.3411053
#> [21,] 0.7960577 0.6271018 0.4271715  0.045656781 0.2831132 0.2649511 0.3749535
#> [22,] 0.6551953 0.4950095 0.4816313  0.074276750 0.3986371 0.4647065 0.5361966
#> [23,] 0.3144140 0.4906541 0.5976951  0.039827912 0.6131496 0.6227135 0.6150362
#> [24,] 0.9505329 0.8869505 0.7912097  0.020898505 0.8324821 0.8803688 0.7571720
#> [25,] 0.7594157 0.7077729 0.5429902 -0.008988908 0.3598931 0.2125928 0.1766070
#>             V33        V34       V35        V36        V37        V38
#>  [1,] 0.6670654 0.63214382 0.6534270 0.54491942 0.39726333 0.40662481
#>  [2,] 0.7363012 0.65354568 0.6584193 0.68860574 0.69786595 0.60030600
#>  [3,] 0.3875195 0.28045568 0.2156820 0.10749914 0.14152945 0.14936418
#>  [4,] 0.4194803 0.20396327 0.1037725 0.29722164 0.40347968 0.37083558
#>  [5,] 0.5831942 0.81470257 0.8035763 0.72545643 0.70743322 0.81751473
#>  [6,] 0.5442004 0.69778013 0.8261484 0.85067476 0.78592733 0.88855213
#>  [7,] 0.3136491 0.32946628 0.4244510 0.14469565 0.13211162 0.29052256
#>  [8,] 0.2185715 0.18261995 0.1721594 0.08807531 0.06949674 0.09361134
#>  [9,] 0.1647801 0.09826355 0.1398078 0.18034824 0.20914791 0.24669260
#> [10,] 0.2702675 0.33185765 0.3914133 0.40941587 0.43647041 0.43456769
#> [11,] 0.3912205 0.36556426 0.2388560 0.22476168 0.13575777 0.22935254
#> [12,] 0.4073785 0.38198293 0.2992131 0.44979403 0.47070588 0.42573776
#> [13,] 0.4409425 0.38828415 0.3011469 0.19026507 0.09762455 0.18367985
#> [14,] 0.6105030 0.64498050 0.5695110 0.44976057 0.31639005 0.25259105
#> [15,] 0.8120720 0.91415941 0.8818604 0.67913534 0.40606316 0.21169280
#> [16,] 0.4685262 0.47495366 0.2384396 0.35426030 0.48394550 0.57818286
#> [17,] 0.9164451 0.88234416 0.7242451 0.51173159 0.32389671 0.19620326
#> [18,] 0.5925982 0.58761611 0.5885751 0.57611082 0.47242351 0.38821089
#> [19,] 0.2559716 0.21126529 0.4791373 0.61985771 0.62476858 0.46121260
#> [20,] 0.2973709 0.27790283 0.1866820 0.08845955 0.18230274 0.14478131
#> [21,] 0.3891740 0.35860483 0.4289646 0.48839833 0.47211144 0.36341867
#> [22,] 0.6296156 0.56223543 0.5588376 0.58279311 0.51309222 0.22358568
#> [23,] 0.6475843 0.65844338 0.6983880 0.75154689 0.77771184 0.83382909
#> [24,] 0.7015000 0.68726994 0.7290675 0.72421846 0.68629764 0.61127036
#> [25,] 0.1892134 0.27192361 0.2402446 0.19460836 0.16603884 0.20441799
#>              V39           V4       V40        V41        V42        V43
#>  [1,] 0.36225654  0.068376104 0.3467173 0.51829545 0.54238006 0.43368204
#>  [2,] 0.56755172  0.036779192 0.6143578 0.56321755 0.44247833 0.48483204
#>  [3,] 0.16106484  0.049364809 0.1976037 0.18960544 0.15279121 0.19072749
#>  [4,] 0.27922932  0.034135873 0.1750898 0.08973225 0.25042879 0.20643329
#>  [5,] 0.76444123  0.167125287 0.6059646 0.45521863 0.47707741 0.45238111
#>  [6,] 0.79824458  0.058938448 0.5570081 0.33135332 0.16593246 0.14190798
#>  [7,] 0.29461326  0.145827805 0.1348183 0.08545275 0.17055308 0.17085354
#>  [8,] 0.10108248  0.035313546 0.1009201 0.06153512 0.09957858 0.18539515
#>  [9,] 0.25485904  0.052146784 0.2620858 0.25576523 0.23079959 0.18762303
#> [10,] 0.38269357  0.068513350 0.3597567 0.35845809 0.41815431 0.37860676
#> [11,] 0.33518977  0.169975095 0.3900785 0.32201154 0.30518583 0.24637297
#> [12,] 0.41116608  0.036431756 0.3198328 0.21505097 0.33436388 0.41245956
#> [13,] 0.16885163  0.042117006 0.1187452 0.15279248 0.17380703 0.16473552
#> [14,] 0.24849415  0.036873382 0.2310044 0.18384672 0.18973012 0.21362036
#> [15,] 0.23368479  0.036017680 0.3354598 0.29560366 0.19314131 0.29958393
#> [16,] 0.64462553  0.017394037 0.4996233 0.21538815 0.19278286 0.14696253
#> [17,] 0.11946768  0.015338842 0.2676750 0.32382158 0.34538575 0.31544237
#> [18,] 0.34936104  0.016367267 0.2842423 0.24422739 0.13891606 0.14764274
#> [19,] 0.28740154  0.072570132 0.3473465 0.35310548 0.30457248 0.24128786
#> [20,] 0.09189244  0.042920654 0.1247936 0.09527317 0.04386406 0.06430727
#> [21,] 0.44721240  0.038825450 0.4299196 0.22921007 0.22276793 0.18301668
#> [22,] 0.07796584  0.055588227 0.4423025 0.53764769 0.65530025 0.53826755
#> [23,] 0.85633852  0.031218051 0.8142433 0.71452655 0.61834714 0.46165192
#> [24,] 0.40264842  0.029331553 0.2481493 0.23235097 0.20523580 0.21723176
#> [25,] 0.26059565 -0.001948201 0.2344058 0.23096665 0.17705729 0.13159897
#>              V44         V45         V46        V47         V48         V49
#>  [1,] 0.48224054 0.522977689  0.44542428 0.29897678 0.187283557 0.089517257
#>  [2,] 0.57538434 0.498987937  0.36690175 0.26411924 0.209243004 0.108326471
#>  [3,] 0.18156306 0.176439086  0.15481390 0.15212908 0.119684196 0.070437894
#>  [4,] 0.10972474 0.055355222  0.02998010 0.05578466 0.068254887 0.045358385
#>  [5,] 0.37512536 0.474335513  0.39104372 0.07566886 0.148110022 0.078961804
#>  [6,] 0.23351767 0.447017081  0.34830806 0.20090875 0.195592815 0.109343914
#>  [7,] 0.08176274 0.082243121  0.08328478 0.09457171 0.093777534 0.058487956
#>  [8,] 0.21167822 0.197923205  0.21007968 0.16780609 0.105222356 0.051965142
#>  [9,] 0.18351006 0.131229726  0.09604597 0.09733522 0.081506500 0.051451249
#> [10,] 0.27670976 0.161489382  0.12326631 0.13042504 0.058383134 0.037617064
#> [11,] 0.20551672 0.239888238  0.20265244 0.12065440 0.127347132 0.070464808
#> [12,] 0.42048016 0.365649192  0.19807079 0.12071947 0.101022080 0.055190412
#> [13,] 0.21872487 0.187342009  0.07935382 0.07608497 0.072900768 0.051007913
#> [14,] 0.13317989 0.041445759  0.03307245 0.03526751 0.033572427 0.009135811
#> [15,] 0.21656801 0.094522553  0.04033216 0.03291637 0.034191276 0.008979556
#> [16,] 0.05202162 0.001884282 -0.01410846 0.05671436 0.066052959 0.016528736
#> [17,] 0.32232831 0.301525800  0.21276090 0.10520326 0.039216294 0.028864915
#> [18,] 0.14781639 0.125223503  0.07550513 0.05440790 0.040502443 0.026693368
#> [19,] 0.20742570 0.196504740  0.12433308 0.13096502 0.113983701 0.070122422
#> [20,] 0.11558270 0.098852713  0.03551800 0.05990365 0.046741789 0.025431741
#> [21,] 0.20951653 0.132863255  0.10179321 0.08065054 0.073927707 0.043572250
#> [22,] 0.29646691 0.154826735  0.09615652 0.04940786 0.065926447 0.018563395
#> [23,] 0.31706464 0.329519149  0.32103916 0.25356114 0.176009718 0.106228775
#> [24,] 0.16342589 0.152524786  0.15227933 0.08725598 0.046895278 0.027093335
#> [25,] 0.14429219 0.135574063  0.11460454 0.03076705 0.006119999 0.015548969
#>               V5         V50          V51         V52         V53         V54
#>  [1,] 0.07086850 0.029909093 0.0318041224 0.026073425 0.015748539 0.014906524
#>  [2,] 0.08174909 0.038772664 0.0232896459 0.014981245 0.006141955 0.010487938
#>  [3,] 0.05135108 0.015379997 0.0205457263 0.015998961 0.006046604 0.011059027
#>  [4,] 0.03689170 0.023247865 0.0140155165 0.011092262 0.002294056 0.009818116
#>  [5,] 0.13369428 0.047607936 0.0196120380 0.029928716 0.039436853 0.029788097
#>  [6,] 0.03420899 0.043882270 0.0278737513 0.012349167 0.009199109 0.018441000
#>  [7,] 0.16763054 0.031285838 0.0248980280 0.027845688 0.018772299 0.016215079
#>  [8,] 0.06630381 0.013976962 0.0201295088 0.014373242 0.006148774 0.004110475
#>  [9,] 0.10708815 0.017778385 0.0189848128 0.013154025 0.013374486 0.016478457
#> [10,] 0.09291002 0.018121576 0.0117256108 0.006207731 0.016568486 0.019251936
#> [11,] 0.14885379 0.017754904 0.0094602247 0.024840989 0.019684982 0.015891461
#> [12,] 0.05687151 0.015231809 0.0156396718 0.011128705 0.009432201 0.010367082
#> [13,] 0.05953424 0.016605221 0.0151741090 0.013569542 0.004198958 0.003237173
#> [14,] 0.04228042 0.009954121 0.0108082204 0.007724844 0.008045104 0.006290874
#> [15,] 0.06836061 0.014997268 0.0145688880 0.014747167 0.020873930 0.015675256
#> [16,] 0.04246584 0.010272395 0.0054254460 0.020387097 0.027912641 0.016726626
#> [17,] 0.02015413 0.013926816 0.0086080193 0.008674688 0.006920301 0.007814285
#> [18,] 0.03304117 0.016852632 0.0036993226 0.006858519 0.004409735 0.004718278
#> [19,] 0.09031776 0.022632201 0.0143866456 0.016459151 0.010477306 0.012976841
#> [20,] 0.06525295 0.013637702 0.0134865571 0.008861401 0.010722848 0.007721564
#> [21,] 0.06067964 0.024050952 0.0150266791 0.008179910 0.008861654 0.019696017
#> [22,] 0.03760286 0.007667397 0.0007216663 0.005248955 0.012807533 0.012227509
#> [23,] 0.05953454 0.041190491 0.0247168620 0.012239881 0.012456929 0.011323170
#> [24,] 0.03230794 0.023371190 0.0082877411 0.010137776 0.002450689 0.005032709
#> [25,] 0.02019576 0.009930404 0.0047181013 0.005670791 0.004791874 0.004151799
#>               V55         V56         V57         V58          V59         V6
#>  [1,] 0.014101853 0.010419469 0.008160942 0.008530946 0.0117256141 0.13201644
#>  [2,] 0.006889141 0.007568956 0.007940251 0.002545478 0.0033890385 0.09392872
#>  [3,] 0.005221602 0.005725691 0.004146372 0.007903850 0.0085590688 0.12255944
#>  [4,] 0.001674430 0.010644756 0.004892123 0.008382423 0.0006778001 0.01245517
#>  [5,] 0.017490249 0.024909751 0.013758708 0.007122528 0.0023634834 0.01859421
#>  [6,] 0.016879762 0.017635378 0.018028850 0.015823326 0.0133424857 0.12469346
#>  [7,] 0.015733810 0.010632426 0.005835096 0.009957653 0.0207050093 0.08359138
#>  [8,] 0.003924018 0.004821730 0.003228240 0.003611159 0.0022526229 0.08710162
#>  [9,] 0.012072445 0.010070662 0.007505039 0.012488025 0.0081390691 0.15749187
#> [10,] 0.005419322 0.014415462 0.014445537 0.010931238 0.0185853079 0.06405476
#> [11,] 0.027257254 0.007870080 0.021689071 0.021314314 0.0192856645 0.11209914
#> [12,] 0.007017303 0.005158056 0.004115720 0.004460089 0.0053901787 0.09422400
#> [13,] 0.003076029 0.003634463 0.001217847 0.004833358 0.0054546969 0.14067161
#> [14,] 0.006241807 0.007495814 0.007628873 0.006020735 0.0074662566 0.08574215
#> [15,] 0.006517812 0.014171321 0.023470765 0.008819494 0.0078436071 0.06467402
#> [16,] 0.023157775 0.004139884 0.002446458 0.005168141 0.0095428875 0.05523726
#> [17,] 0.002521805 0.001392230 0.002054886 0.002878195 0.0059656640 0.01770351
#> [18,] 0.005365254 0.002752061 0.004889657 0.004385859 0.0048117221 0.08626215
#> [19,] 0.009136532 0.007127755 0.012526006 0.009153637 0.0097884009 0.13597492
#> [20,] 0.004058949 0.008593343 0.005660661 0.003019972 0.0042587632 0.08076554
#> [21,] 0.014679391 0.011054776 0.014402987 0.014645015 0.0115769445 0.06224374
#> [22,] 0.010019060 0.018383290 0.012168476 0.003519083 0.0077351405 0.09710631
#> [23,] 0.012724137 0.010475882 0.005329147 0.011200496 0.0100033112 0.09645612
#> [24,] 0.008130677 0.004240927 0.011628614 0.011796862 0.0070983254 0.07353956
#> [25,] 0.002804518 0.001822486 0.001653745 0.004907680 0.0021035533 0.01018151
#>               V60         V7         V8         V9
#>  [1,] 0.008208630 0.12781762 0.13407550 0.24450703
#>  [2,] 0.003121204 0.13118602 0.15737847 0.26448375
#>  [3,] 0.005732292 0.14990746 0.14863885 0.21840493
#>  [4,] 0.008113486 0.04442083 0.08138778 0.22374471
#>  [5,] 0.010144855 0.07386261 0.26566839 0.56652872
#>  [6,] 0.009230081 0.21553217 0.33715235 0.50986965
#>  [7,] 0.012709888 0.09045028 0.04445545 0.10779862
#>  [8,] 0.004152880 0.12043579 0.13800718 0.19637503
#>  [9,] 0.005571647 0.16603439 0.16700374 0.21412227
#> [10,] 0.012133869 0.07874922 0.06819204 0.03633465
#> [11,] 0.012874185 0.13564893 0.20122622 0.11704382
#> [12,] 0.003719391 0.10725763 0.13442078 0.26083505
#> [13,] 0.001383760 0.15096124 0.11611216 0.15765078
#> [14,] 0.005327655 0.10253589 0.11173358 0.13292318
#> [15,] 0.013189577 0.09890271 0.08720810 0.02082112
#> [16,] 0.009305404 0.03050580 0.06875216 0.14051231
#> [17,] 0.003549875 0.06342481 0.04127016 0.05939974
#> [18,] 0.003157425 0.07634209 0.09630438 0.09672312
#> [19,] 0.007797494 0.18019521 0.22361085 0.28261941
#> [20,] 0.003115569 0.12775468 0.09844260 0.11066697
#> [21,] 0.005831905 0.09243630 0.07128431 0.12926930
#> [22,] 0.014914155 0.09126668 0.15839625 0.01044638
#> [23,] 0.010810268 0.14851018 0.15258811 0.04983932
#> [24,] 0.009166487 0.11488896 0.11912904 0.07404111
#> [25,] 0.004143153 0.04604118 0.08518998 0.10486989
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913