Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11         V12        V13        V14
#>  [1,] 0.055945449  0.37941742 0.42930323  0.29084830 0.55130949 0.91949274
#>  [2,] 0.089446277  0.53223436 0.37493314  0.15475418 0.17653298 0.30296134
#>  [3,] 0.002967700  0.21948267 0.22892038  0.16892231 0.05409512 0.15107443
#>  [4,] 0.040003659  0.43774738 0.56711684  0.57862921 0.66785126 0.66944808
#>  [5,] 0.038417778  0.46055606 0.50452547  0.46836704 0.46452500 0.39150605
#>  [6,] 0.039966221  0.26674171 0.29537631  0.31606300 0.24051559 0.17452436
#>  [7,] 0.028714357  0.27383085 0.31330405  0.38336322 0.39495220 0.41379560
#>  [8,] 0.015899509  0.02644434 0.08572119  0.16435342 0.24054528 0.25990197
#>  [9,] 0.017803368  0.31055959 0.31379169  0.35497579 0.32531109 0.24233548
#> [10,] 0.032979532  0.19683358 0.24612901  0.28493684 0.28792941 0.23652449
#> [11,] 0.108300000  0.59660000 0.53040000  0.22510000 0.24020000 0.26890000
#> [12,] 0.034508446  0.08997138 0.22782544  0.34406845 0.36512726 0.41754801
#> [13,] 0.012613621  0.26836108 0.34629576  0.32289595 0.26787906 0.20888147
#> [14,] 0.005415662  0.22404236 0.03157641 -0.10237349 0.15949051 0.30568687
#> [15,] 0.007690627  0.07162935 0.03185303  0.15852307 0.12828424 0.08900129
#> [16,] 0.023490570 -0.03058945 0.08112610  0.36316559 0.33860790 0.32139031
#> [17,] 0.022606365  0.06181921 0.08877004  0.09764314 0.13085802 0.15796551
#> [18,] 0.018087682  0.18756238 0.16607645  0.22161743 0.32442548 0.14333251
#> [19,] 0.006369727  0.10990723 0.09370000  0.08457773 0.09505019 0.09393457
#> [20,] 0.009257122  0.13167931 0.11976039  0.06898010 0.09315380 0.12838192
#> [21,] 0.012413932  0.13320974 0.11819407  0.06522052 0.03599249 0.08618094
#> [22,] 0.015821169  0.10484107 0.11084234  0.09519090 0.10329075 0.12309515
#> [23,] 0.008530072  0.10777849 0.10450058  0.15397311 0.17951089 0.20085493
#> [24,] 0.026901515  0.10057982 0.06381028  0.16911525 0.15419050 0.08740509
#> [25,] 0.035763913  0.25857358 0.28589968  0.32090716 0.49575338 0.60747850
#>              V15        V16        V17       V18        V19            V2
#>  [1,] 0.89269346 0.85452990 0.73278590 0.8125495 0.92034360  0.0583639764
#>  [2,] 0.41350181 0.46387636 0.25751191 0.2234507 0.29595581  0.0866855924
#>  [3,] 0.28035702 0.58527542 0.72536749 0.8103418 0.85457255  0.0089895108
#>  [4,] 0.69767935 0.72493619 0.87916753 0.9461673 0.90115419  0.0244435374
#>  [5,] 0.26543326 0.30517179 0.33924114 0.3507221 0.51907088  0.0471129406
#>  [6,] 0.13862155 0.16055597 0.20769424 0.3266088 0.50595560  0.0488085217
#>  [7,] 0.43396598 0.46773857 0.59173846 0.6778013 0.70371356  0.0324418787
#>  [8,] 0.30087993 0.48349384 0.59650963 0.6746439 0.85572478 -0.0037798473
#>  [9,] 0.16288455 0.17750998 0.17853416 0.2681787 0.41876898  0.0186173848
#> [10,] 0.20865984 0.24169806 0.26171351 0.2290093 0.30392050  0.0585066540
#> [11,] 0.66460000 0.66320000 0.16740000 0.0837000 0.43310000  0.1070000000
#> [12,] 0.56743346 0.66866985 0.80414133 0.8243423 0.92996199  0.0381479256
#> [13,] 0.16439936 0.17477504 0.25186643 0.3894804 0.60723098  0.0693803968
#> [14,] 0.45702986 0.64451486 0.65636426 0.7099021 0.87643196 -0.0040552481
#> [15,] 0.11737764 0.12388550 0.09756057 0.1527531 0.08452897  0.0112467121
#> [16,] 0.05135994 0.02571141 0.15591141 0.4314371 0.26233823  0.0209607436
#> [17,] 0.17829176 0.27462374 0.17881866 0.2671149 0.49414712  0.0411352546
#> [18,] 0.22662378 0.26871102 0.30596830 0.3366528 0.28469370  0.0497187241
#> [19,] 0.15031269 0.16689648 0.12773887 0.1390801 0.11111718  0.0120023441
#> [20,] 0.17561976 0.16589195 0.16962185 0.1728503 0.35375266  0.0147442912
#> [21,] 0.08189459 0.12027990 0.24285288 0.2925494 0.36496590 -0.0003526983
#> [22,] 0.14392566 0.22397111 0.33341316 0.3262871 0.29967561  0.0156724820
#> [23,] 0.22489633 0.23745679 0.19035898 0.1995391 0.16413501  0.0127943045
#> [24,] 0.16578193 0.22451899 0.17395723 0.1240643 0.16721316  0.0389422313
#> [25,] 0.70448553 0.69168587 0.48257561 0.3452434 0.27568738  0.0704337523
#>             V20       V21       V22        V23         V24        V25       V26
#>  [1,] 0.8992307 0.9001051 0.9083072 0.96809253 0.862912676 0.68185310 0.6920744
#>  [2,] 0.8522480 0.9913707 0.4081999 0.09881571 0.175676837 0.19781569 0.3548774
#>  [3,] 0.7170948 0.7530876 0.9302740 0.90833233 0.450389967 0.14284479 0.2844248
#>  [4,] 0.8349118 0.6985197 0.3909287 0.05182247 0.004942589 0.24890859 0.3991826
#>  [5,] 0.7004930 0.7380053 0.7782689 0.77425906 0.789978678 0.77140389 0.8284423
#>  [6,] 0.6536594 0.7961093 0.8387978 0.81815698 0.943109306 0.97649827 1.0286133
#>  [7,] 0.7764587 0.8421221 0.8861819 0.95518755 1.010416978 1.01105240 1.0126681
#>  [8,] 0.9660254 0.9469127 0.9927909 1.06161034 0.953163596 0.90323621 0.8293952
#>  [9,] 0.5091040 0.6696220 0.7745899 0.78867619 0.895427343 0.94073464 0.9923695
#> [10,] 0.3555905 0.4727828 0.6194021 0.68049211 0.685138885 0.66185314 0.6694553
#> [11,] 0.8718000 0.7992000 0.3712000 0.17030000 0.161100000 0.20860000 0.2847000
#> [12,] 0.9910480 0.9382165 0.8719546 0.77486892 0.719743287 0.56589669 0.4115631
#> [13,] 0.7019038 0.8057630 0.8191075 0.83902794 0.974621886 0.98944032 0.9661578
#> [14,] 0.9427592 1.0377790 1.0370295 0.95188631 0.630257436 0.39076591 0.4192855
#> [15,] 0.1669732 0.2722918 0.1298508 0.08704984 0.115199739 0.01868457 0.1631853
#> [16,] 0.5420215 0.6889251 0.6937343 0.73296460 0.806201863 0.99019508 0.8716063
#> [17,] 0.2454395 0.4700470 0.5076852 0.39351527 0.564234449 0.48580203 0.5698394
#> [18,] 0.1264932 0.1702571 0.1299859 0.17279947 0.176894327 0.24771593 0.2594056
#> [19,] 0.1299162 0.2150986 0.2847941 0.41989180 0.615272659 0.81821836 0.9638980
#> [20,] 0.4424799 0.4336761 0.4758256 0.68581463 0.814619630 0.82582871 0.7693601
#> [21,] 0.5370165 0.6479720 0.8025080 0.94832353 0.890210394 0.88392177 0.9705766
#> [22,] 0.3149383 0.4214080 0.5772148 0.64951386 0.704900294 0.74719680 0.8052018
#> [23,] 0.1949713 0.2952302 0.2756747 0.37281714 0.553361659 0.74953979 0.8406704
#> [24,] 0.2465659 0.2713117 0.1844147 0.26667434 0.221996330 0.03996521 0.2377523
#> [25,] 0.4224417 0.5012139 0.4177211 0.33613337 0.609456071 0.62932256 0.6634812
#>             V27       V28       V29           V3       V30       V31        V32
#>  [1,] 0.5781767 0.4687113 0.4440734  0.046950975 0.2936938 0.2339427 0.09855438
#>  [2,] 0.2506042 0.3754878 0.4459289  0.070592531 0.6159951 0.2668721 0.33189763
#>  [3,] 0.6326371 0.8611915 0.9793401 -0.013629925 0.8789535 0.4962577 0.43078207
#>  [4,] 0.4344491 0.3974992 0.1994353  0.078118787 0.1254948 0.2356284 0.26237293
#>  [5,] 0.8713503 0.8159090 0.6540660  0.053433293 0.5409583 0.4830838 0.51263729
#>  [6,] 1.0092653 0.9289726 0.8375370  0.048363195 0.6810901 0.4899944 0.34054055
#>  [7,] 0.9778597 0.9319412 0.8275953  0.037150104 0.6584802 0.4588134 0.33205338
#>  [8,] 0.7826612 0.7887814 0.6545673  0.007454512 0.4559028 0.2179077 0.13167006
#>  [9,] 1.0219400 1.0132540 0.8861493  0.037363850 0.6615534 0.5378479 0.41872623
#> [10,] 0.7317086 0.7786224 0.7623068  0.064786786 0.7558233 0.7416763 0.67321560
#> [11,] 0.2211000 0.6134000 0.5807000  0.025700000 0.6925000 0.3825000 0.43030000
#> [12,] 0.3593044 0.2427657 0.3006671  0.055165445 0.3610212 0.2121475 0.24663529
#> [13,] 0.9941436 0.9986432 0.8875720  0.035933368 0.6506020 0.3892593 0.30860818
#> [14,] 0.5839206 0.7874669 0.7421709  0.024885637 0.4977436 0.2247805 0.33353440
#> [15,] 0.3638226 0.5280828 0.5595723  0.021183717 0.5790153 0.6478354 0.58929625
#> [16,] 0.6839830 0.5310813 0.5324513  0.077663070 0.3789172 0.4128858 0.48401347
#> [17,] 0.4148911 0.3427699 0.6090908  0.041638057 0.8068171 0.7913377 0.45489181
#> [18,] 0.1318294 0.5254415 0.7104581  0.037768870 0.6462434 0.5219727 0.62999024
#> [19,] 0.8287951 0.5852297 0.4707295  0.006907812 0.3556340 0.2749730 0.14389296
#> [20,] 0.7566616 0.6745626 0.4612937  0.018519798 0.3063256 0.2683794 0.31355788
#> [21,] 0.9756299 0.9220758 0.8379289  0.003190724 0.7213894 0.6140840 0.57762774
#> [22,] 0.8023120 0.8228398 0.8390785  0.024944517 0.7823122 0.6934495 0.65941118
#> [23,] 0.7499178 0.7477434 0.7308502  0.014641912 0.5407023 0.4028108 0.29556523
#> [24,] 0.5911484 0.8213222 0.8732602  0.024688749 0.8249941 0.7007823 0.46633754
#> [25,] 0.6774096 0.8079057 0.7597633  0.048116518 0.7101102 0.6803228 0.51044938
#>              V33        V34        V35         V36        V37        V38
#>  [1,] 0.04478024 0.12884918 0.18373791  0.17049309 0.11841922 0.19341887
#>  [2,] 0.52064020 0.67688793 0.75446617  0.78524439 0.75425068 0.83199417
#>  [3,] 0.40851137 0.40766028 0.29330928  0.15950361 0.19757157 0.48497091
#>  [4,] 0.30902662 0.27453694 0.37210840  0.01404400 0.12948148 0.16853991
#>  [5,] 0.50920063 0.41178367 0.21794677  0.27642658 0.29113720 0.37657662
#>  [6,] 0.22794920 0.24723353 0.17395995  0.02048941 0.08742464 0.15823749
#>  [7,] 0.18067193 0.05965752 0.05916334  0.11158416 0.20680203 0.27902052
#>  [8,] 0.16197189 0.26115595 0.38019630  0.39589820 0.30210910 0.20873409
#>  [9,] 0.37431544 0.20056258 0.07537603 -0.02930562 0.04860607 0.08669695
#> [10,] 0.55312557 0.51396592 0.43163452  0.35461559 0.28246764 0.27047982
#> [11,] 0.77910000 0.87030000 1.00000000  0.92120000 0.93860000 0.93030000
#> [12,] 0.37782650 0.41203869 0.36075008  0.31537742 0.37149910 0.40861658
#> [13,] 0.24868712 0.22331886 0.20629111  0.07257257 0.02196079 0.05524169
#> [14,] 0.36814401 0.16831967 0.12169593  0.09446326 0.04484083 0.04169738
#> [15,] 0.65979984 0.89470950 1.00584483  0.96787284 0.83520946 0.69105449
#> [16,] 0.60480144 0.52542528 0.54429596  0.58782930 0.55507949 0.22388428
#> [17,] 0.37157966 0.48340577 0.63694304  0.66647385 0.59374571 0.73319915
#> [18,] 0.71365570 0.63365996 0.59735288  0.76166149 0.87409829 0.93603962
#> [19,] 0.18655605 0.21782910 0.23953535  0.17493925 0.05982539 0.06267051
#> [20,] 0.27887197 0.32307463 0.25717648  0.28002876 0.37049806 0.32009951
#> [21,] 0.56983861 0.60376582 0.61849638  0.56341789 0.49829416 0.37131864
#> [22,] 0.70016343 0.72716319 0.71250997  0.64186345 0.50837001 0.42475641
#> [23,] 0.30677290 0.32845163 0.33618622  0.21522256 0.08804450 0.11426041
#> [24,] 0.64464401 0.88134975 0.85886533  0.80945981 0.63847827 0.54028347
#> [25,] 0.20003902 0.25081210 0.44415987  0.46252837 0.23087428 0.25481236
#>              V39          V4        V40         V41        V42        V43
#>  [1,] 0.33675626 0.076479486 0.31419671 0.208297007 0.12270969 0.26884332
#>  [2,] 0.73835337 0.074798643 0.54719002 0.326663202 0.20476535 0.20822786
#>  [3,] 0.49289260 0.005519145 0.36888698 0.098498797 0.14621480 0.44638612
#>  [4,] 0.14914044 0.138855903 0.09367753 0.029292083 0.12276651 0.14781469
#>  [5,] 0.51363434 0.050093993 0.40193608 0.263405701 0.40335913 0.42187389
#>  [6,] 0.10143256 0.030680122 0.12506316 0.008619786 0.11819161 0.24711394
#>  [7,] 0.28317345 0.045830055 0.32176764 0.310468806 0.29789147 0.24952535
#>  [8,] 0.36166902 0.109207632 0.43279588 0.459550507 0.44665551 0.37333785
#>  [9,] 0.10042462 0.053715500 0.11736063 0.063061437 0.09080731 0.19501425
#> [10,] 0.26634704 0.067102943 0.35657160 0.504788413 0.47962795 0.35802705
#> [11,] 0.73140000 0.083700000 0.47910000 0.208700000 0.20160000 0.16690000
#> [12,] 0.36466859 0.069454081 0.34259019 0.303979252 0.31442597 0.28214682
#> [13,] 0.13450849 0.019330660 0.14017879 0.018084193 0.02823725 0.14195845
#> [14,] 0.06220399 0.072596412 0.10935783 0.258222020 0.28655539 0.11217677
#> [15,] 0.65085572 0.039219966 0.63155406 0.512079304 0.43078283 0.29296240
#> [16,] 0.01967587 0.049760590 0.40832440 0.487701508 0.64953663 0.53772284
#> [17,] 0.91112592 0.020059525 0.91155529 0.754858546 0.69926782 0.71890495
#> [18,] 0.95704615 0.030786860 0.89729707 0.867734691 0.76192866 0.54576937
#> [19,] 0.18164083 0.012053318 0.24303028 0.296760354 0.30170118 0.22618828
#> [20,] 0.36219428 0.002654080 0.30377764 0.220611272 0.17799662 0.16713230
#> [21,] 0.16603792 0.039884986 0.13577779 0.149912405 0.05223034 0.04554749
#> [22,] 0.31461765 0.026570469 0.24483474 0.192244202 0.17575371 0.20297682
#> [23,] 0.14595446 0.030252786 0.14990136 0.217083698 0.23208997 0.21330682
#> [24,] 0.45808699 0.016632067 0.59059518 0.528787691 0.45702055 0.26873792
#> [25,] 0.26773511 0.029864011 0.31900923 0.255990725 0.14929044 0.18229517
#>              V44        V45        V46        V47        V48         V49
#>  [1,] 0.25490543 0.11195633 0.08040368 0.09910388 0.09342718 0.084252632
#>  [2,] 0.24523874 0.43552305 0.33144268 0.15430456 0.16368639 0.103704913
#>  [3,] 0.37972775 0.10400057 0.18060759 0.16236820 0.08226185 0.064959847
#>  [4,] 0.06437885 0.09790797 0.05749697 0.04485479 0.06216989 0.037822381
#>  [5,] 0.42482649 0.42154696 0.23333046 0.14894591 0.16230856 0.089320223
#>  [6,] 0.26568488 0.21984033 0.23643397 0.17665108 0.08077930 0.056716562
#>  [7,] 0.20555024 0.14446099 0.12320741 0.14474392 0.10329066 0.057011418
#>  [8,] 0.26658152 0.19125811 0.11632751 0.12654524 0.05157616 0.023693398
#>  [9,] 0.27139629 0.24667551 0.19123332 0.16959767 0.13480231 0.068854349
#> [10,] 0.27136651 0.34591696 0.30443317 0.21510159 0.14924150 0.087761830
#> [11,] 0.28720000 0.43740000 0.30970000 0.15780000 0.05530000 0.033400000
#> [12,] 0.20055477 0.10851873 0.05769052 0.06011751 0.03569985 0.017808893
#> [13,] 0.19286575 0.17277091 0.18354187 0.13933986 0.08348168 0.037963378
#> [14,] 0.21389927 0.22913023 0.17717893 0.10252338 0.03384481 0.031608648
#> [15,] 0.17026248 0.13093939 0.07964799 0.08371186 0.08486343 0.039043994
#> [16,] 0.34448622 0.15808983 0.07684092 0.03782627 0.06537370 0.008621892
#> [17,] 0.72436224 0.58841213 0.44531355 0.30726303 0.18218255 0.092134227
#> [18,] 0.29540767 0.39845736 0.48851709 0.38926329 0.26576306 0.177768463
#> [19,] 0.14561016 0.10943184 0.11684317 0.05978028 0.03375820 0.027869531
#> [20,] 0.18471020 0.12382233 0.08623596 0.01761982 0.01671628 0.021276327
#> [21,] 0.02125937 0.03448755 0.01134936 0.01120165 0.02149105 0.007549328
#> [22,] 0.17184147 0.10985410 0.07174192 0.05049977 0.05113753 0.022126558
#> [23,] 0.12863464 0.06831019 0.07821010 0.03641252 0.03400691 0.026693331
#> [24,] 0.10219159 0.09897240 0.06967882 0.09321918 0.04953105 0.023751110
#> [25,] 0.13737980 0.10040326 0.09251590 0.06790914 0.02687564 0.020656141
#>               V5          V50          V51         V52         V53
#>  [1,] 0.10559991  0.026963351  0.031457971 0.008276214 0.014963544
#>  [2,] 0.05923000  0.046471964  0.023941874 0.012598809 0.012736188
#>  [3,] 0.02323726  0.027081829  0.010952498 0.019662439 0.009730200
#>  [4,] 0.08398645  0.026304240  0.033910582 0.028125534 0.028651925
#>  [5,] 0.09023785  0.027861888  0.025748043 0.015547028 0.011480601
#>  [6,] 0.05793150  0.017754642  0.021304954 0.014506484 0.002554204
#>  [7,] 0.11373959  0.029599877  0.018480940 0.013790689 0.021172442
#>  [8,] 0.17891093  0.020139113  0.013207663 0.014016669 0.013869284
#>  [9,] 0.09430677  0.022326647  0.019913826 0.012235085 0.008749182
#> [10,] 0.07621968  0.025823309  0.023579274 0.025136994 0.013592978
#> [11,] 0.07480000  0.020900000  0.017200000 0.018000000 0.011000000
#> [12,] 0.10418596  0.005170298  0.005435087 0.007653446 0.011110716
#> [13,] 0.01814610  0.004754121  0.020343271 0.013822669 0.008579145
#> [14,] 0.12539084 -0.001100076  0.002359113 0.008648329 0.002352025
#> [15,] 0.01706680  0.021985522  0.020523800 0.004059769 0.007899539
#> [16,] 0.05605878  0.004161081 -0.001988195 0.002283167 0.013791152
#> [17,] 0.11350036  0.021806331  0.020920435 0.010020621 0.013636150
#> [18,] 0.06238865  0.070016639  0.035817726 0.016836273 0.015494150
#> [19,] 0.02310605  0.009797852  0.002551562 0.007124609 0.002634765
#> [20,] 0.01695427  0.005946108  0.001551683 0.008441515 0.006632390
#> [21,] 0.05615093  0.009636161  0.006040172 0.006738959 0.005940285
#> [22,] 0.04951159  0.017221704  0.010025515 0.007192131 0.008353405
#> [23,] 0.03099385  0.010741913  0.009121665 0.011285688 0.007981476
#> [24,] 0.02371776  0.020525052  0.012735823 0.009800886 0.020295705
#> [25,] 0.13217842  0.015681566  0.010550908 0.014383090 0.007544866
#>                 V54          V55          V56          V57          V58
#>  [1,]  0.0152213285  0.029202527 0.0088689932  0.010551908 0.0240604780
#>  [2,]  0.0253971894  0.017872760 0.0204287706  0.015169690 0.0166186298
#>  [3,]  0.0018577193  0.002998808 0.0060919737 -0.001935124 0.0040525617
#>  [4,]  0.0121098177  0.019561896 0.0180662397  0.004866441 0.0370297340
#>  [5,]  0.0131994319  0.015313824 0.0111264233  0.009283405 0.0123108102
#>  [6,]  0.0006523226  0.003570473 0.0013259076  0.002558451 0.0011615457
#>  [7,]  0.0136920938  0.006739225 0.0128357751  0.009990032 0.0141804459
#>  [8,]  0.0070764615  0.001105878 0.0104354105  0.008328763 0.0113496367
#>  [9,]  0.0080521988  0.001667946 0.0033229394  0.003713683 0.0003717151
#> [10,]  0.0121043992  0.010140159 0.0088852225  0.008226588 0.0092063516
#> [11,]  0.0234000000  0.027600000 0.0032000000  0.008400000 0.0122000000
#> [12,]  0.0167221165  0.011191607 0.0180130657  0.009094475 0.0102575401
#> [13,]  0.0080190333  0.002974283 0.0029111595  0.003514805 0.0034025887
#> [14,] -0.0038312786 -0.005709925 0.0003210768  0.000929963 0.0074521566
#> [15,]  0.0097822052  0.011482157 0.0088835329  0.006240150 0.0067080901
#> [16,]  0.0149593334  0.009326826 0.0195274104  0.014469391 0.0032481970
#> [17,]  0.0101935356  0.012635242 0.0181949532  0.002325538 0.0177837395
#> [18,]  0.0104566139  0.017794415 0.0122624635  0.002827543 0.0160289829
#> [19,]  0.0027359374  0.005128125 0.0034562502  0.002564063 0.0036583983
#> [20,]  0.0083430679  0.005688716 0.0018061370  0.004582662 0.0090384293
#> [21,]  0.0019271357  0.001534362 0.0042250575  0.003501003 0.0008915616
#> [22,]  0.0102394462  0.007134087 0.0085394188  0.010209442 0.0058930647
#> [23,]  0.0047825185  0.007397115 0.0029606179  0.005890968 0.0047054062
#> [24,]  0.0174802722  0.007611892 0.0099366564  0.008771179 0.0059295503
#> [25,]  0.0086074376  0.005938799 0.0090920855  0.005526262 0.0044535520
#>                V59         V6          V60         V7           V8           V9
#>  [1,] 0.0131857053 0.17654331 0.0065857445 0.27073377  0.307674215  0.304393897
#>  [2,] 0.0155172945 0.10966167 0.0072020612 0.14492275  0.367887445  0.548408922
#>  [3,] 0.0017873741 0.05245569 0.0051542627 0.06249735 -0.002906999  0.104909304
#>  [4,] 0.0151798813 0.13795061 0.0169398649 0.19242236  0.126365172  0.275603716
#>  [5,] 0.0089019736 0.11440610 0.0054937539 0.13797273  0.190387487  0.328241755
#>  [6,] 0.0025802441 0.02594104 0.0056229660 0.08987256  0.133068986  0.192454192
#>  [7,] 0.0080956038 0.15641262 0.0066782073 0.14320517  0.176993032  0.242163885
#>  [8,] 0.0178023154 0.02985629 0.0145288399 0.06011726  0.069096807  0.010555464
#>  [9,] 0.0033071751 0.09353447 0.0046150599 0.08347629  0.158697840  0.230076279
#> [10,] 0.0093073895 0.13811301 0.0071675328 0.10869037  0.114237635  0.188915160
#> [11,] 0.0082000000 0.11250000 0.0143000000 0.33220000  0.459000000  0.552600000
#> [12,] 0.0136785513 0.09994205 0.0079666836 0.09412079  0.098141062  0.079485114
#> [13,] 0.0042836783 0.07466006 0.0045212045 0.13104579  0.138056949  0.204687445
#> [14,] 0.0019900724 0.14821057 0.0001101923 0.16348450  0.227508998  0.274153885
#> [15,] 0.0120342910 0.06171874 0.0041997168 0.05215053  0.067954828  0.009337117
#> [16,] 0.0061540971 0.11676069 0.0157072084 0.11118786  0.133866540 -0.003545781
#> [17,] 0.0220508341 0.11619346 0.0162246412 0.15123938  0.206018916  0.079703370
#> [18,] 0.0106434744 0.10326561 0.0126029690 0.21322957  0.220313201  0.040511576
#> [19,] 0.0006808592 0.03250136 0.0034146486 0.05591640  0.059525973  0.106231053
#> [20,] 0.0051105388 0.01018565 0.0040223460 0.03654942  0.074943513  0.070033726
#> [21,] 0.0011859882 0.09025864 0.0017379366 0.09766392  0.089915123  0.163490723
#> [22,] 0.0061735562 0.07716219 0.0065602020 0.10381890  0.094188225  0.106437910
#> [23,] 0.0034859499 0.09174569 0.0040813440 0.11080835  0.102974336  0.121719796
#> [24,] 0.0070964403 0.05242830 0.0056069070 0.03068105  0.085230271  0.082105681
#> [25,] 0.0081011266 0.27181296 0.0057929818 0.28259596  0.299078484  0.315408441
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3478261