Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.026401550  0.35321466 0.43924258 0.47521525 0.45829495 0.37449497
#>  [2,] 0.012396756  0.12688056 0.24403024 0.23149256 0.25187603 0.14842992
#>  [3,] 0.004256015  0.14512012 0.20948474 0.25288720 0.22947999 0.13598486
#>  [4,] 0.011503529  0.32591410 0.30526966 0.31750023 0.25220945 0.16190161
#>  [5,] 0.050456580 -0.06730889 0.05432947 0.16448011 0.20525003 0.31505772
#>  [6,] 0.040313851 -0.07769414 0.09934947 0.25201653 0.32133445 0.39598525
#>  [7,] 0.030277966  0.27896875 0.38469084 0.49469059 0.49249793 0.48903392
#>  [8,] 0.068437737  0.36106815 0.52129045 0.45272409 0.44924179 0.44283466
#>  [9,] 0.025793479  0.07388064 0.11502945 0.09588067 0.07016464 0.16966575
#> [10,] 0.026853033  0.22010394 0.27110858 0.30574479 0.27661190 0.22722673
#> [11,] 0.054731053  0.20878332 0.22530621 0.24213687 0.30551033 0.33703458
#> [12,] 0.058861717  0.34571266 0.28544353 0.15367102 0.14623337 0.21086937
#> [13,] 0.037550943  0.71773703 0.74276352 0.50625520 0.31849193 0.21964072
#> [14,] 0.067246286  0.26575807 0.37366620 0.36231387 0.51268233 0.64799299
#> [15,] 0.011718523  0.08905649 0.07612286 0.17939684 0.17408832 0.12218437
#> [16,] 0.023825446  0.08255749 0.06677295 0.07367995 0.01510433 0.10749651
#> [17,] 0.033764737  0.29446423 0.37697915 0.48283713 0.45159102 0.49718628
#> [18,] 0.016968821  0.40583028 0.29808711 0.10692414 0.31172419 0.55621354
#> [19,] 0.014622920  0.08998277 0.07670350 0.04745835 0.07903492 0.08519559
#> [20,] 0.012466197  0.18041498 0.13637571 0.07993528 0.08201317 0.08876678
#> [21,] 0.012535270  0.11218932 0.09856917 0.09026300 0.07425636 0.13072515
#> [22,] 0.001280858  0.20203199 0.18322169 0.23907160 0.38463438 0.30325936
#> [23,] 0.015734609  0.07100317 0.07886386 0.09796671 0.13266526 0.14684466
#> [24,] 0.021004121  0.11556125 0.14073112 0.10033863 0.09816098 0.12424151
#> [25,] 0.012959665  0.09066340 0.12439365 0.17813251 0.24459508 0.27633086
#>               V15        V16       V17       V18       V19           V2
#>  [1,] 0.311027925 0.36495594 0.4371567 0.5124432 0.5998943  0.031168429
#>  [2,] 0.067202233 0.10805648 0.2357701 0.3269239 0.4526147  0.016742453
#>  [3,] 0.024755809 0.18501030 0.3531259 0.4517393 0.5427333  0.017040009
#>  [4,] 0.155864278 0.22507343 0.2202418 0.2356662 0.1876538  0.018854203
#>  [5,] 0.383393377 0.50565678 0.6322596 0.7313868 0.9054147  0.053471553
#>  [6,] 0.412891406 0.53921540 0.6925313 0.7687053 0.8190233  0.040749456
#>  [7,] 0.555332069 0.72781519 0.8178836 0.9181643 1.0031397  0.052300000
#>  [8,] 0.610806240 0.81403489 0.9559024 0.9457249 0.8039454  0.129041500
#>  [9,] 0.198685684 0.37090702 0.4624800 0.6519927 0.8044399 -0.002541770
#> [10,] 0.200216607 0.18650184 0.2005000 0.2600551 0.3422910  0.034106051
#> [11,] 0.352804673 0.41064152 0.4127787 0.2797741 0.3961715  0.078860121
#> [12,] 0.286948059 0.31667327 0.1988278 0.2484747 0.3728254  0.062793427
#> [13,] 0.262831020 0.31169489 0.2310719 0.2802082 0.4080630  0.076549369
#> [14,] 0.697825040 0.67210596 0.7350167 0.7885236 0.9039327  0.064804609
#> [15,] 0.149090495 0.19102398 0.1508515 0.2061045 0.2361147  0.032830689
#> [16,] 0.072410109 0.17682400 0.1542474 0.2755252 0.3737184  0.036043509
#> [17,] 0.640566621 0.71732329 0.8813540 0.9304472 0.8772173  0.045628145
#> [18,] 0.674682762 0.92426122 0.9703302 0.8938831 0.6024057  0.066575397
#> [19,] 0.006987255 0.10160914 0.1579871 0.4079438 0.3427208  0.011267290
#> [20,] 0.121090218 0.22475805 0.2792238 0.2788029 0.3908051  0.002620358
#> [21,] 0.068102524 0.07120975 0.2092676 0.3211082 0.3167936  0.021951050
#> [22,] 0.307358434 0.54814478 0.4129190 0.1566844 0.2767294 -0.002551897
#> [23,] 0.117100354 0.11348487 0.1518883 0.1887346 0.2385616  0.025961382
#> [24,] 0.127416039 0.19164881 0.2979133 0.3178474 0.2805561  0.014015486
#> [25,] 0.290577545 0.28194906 0.2059075 0.2482503 0.3220551  0.007801688
#>             V20       V21       V22         V23         V24        V25
#>  [1,] 0.6946396 0.8059595 0.8819368  0.85396812  0.87145599 0.89879922
#>  [2,] 0.5247218 0.6526008 0.7826791  0.79606636  0.77538038 0.86238136
#>  [3,] 0.6180860 0.7237614 0.8034245  0.82545848  0.80650915 0.86608589
#>  [4,] 0.1285377 0.1089684 0.2213468  0.40590633  0.55566611 0.68720103
#>  [5,] 0.9516996 0.8570159 0.8683516  0.87545375  0.83553736 0.70327838
#>  [6,] 0.8838715 0.8107040 0.7809909  0.88388465  1.03748538 0.96925460
#>  [7,] 0.9971501 0.8327529 0.6568705  0.60384778  0.48894033 0.31471094
#>  [8,] 0.7285975 0.7606211 0.4051248 -0.00344946 -0.02003459 0.24344358
#>  [9,] 0.8733659 0.8808844 0.9706849  1.04557602  1.00031612 0.97260520
#> [10,] 0.4129491 0.5083498 0.5561445  0.61525475  0.70972816 0.77046615
#> [11,] 0.4934876 0.5377391 0.5381711  0.52060088  0.53909555 0.56442575
#> [12,] 0.7045967 0.7731518 0.5018399  0.33367554  0.29499405 0.36626314
#> [13,] 0.5576607 0.6951042 0.8584591  0.95652620  0.99411746 0.71423187
#> [14,] 0.9055822 0.9753613 0.9969805  0.92599137  0.73569455 0.55726780
#> [15,] 0.1452141 0.2179466 0.1750025  0.25691412  0.27726876 0.17897200
#> [16,] 0.3639078 0.3079838 0.1209009  0.12118155  0.28929752 0.14371894
#> [17,] 0.8116882 0.7330725 0.7121220  0.56126665  0.53232843 0.61115419
#> [18,] 0.4481387 0.5095446 0.5512716  0.73014891  0.86481183 0.81561678
#> [19,] 0.3014736 0.1987752 0.1970451  0.18354651  0.20000048 0.01956733
#> [20,] 0.5752243 0.6732466 0.8814197  1.01779395  0.88584037 0.82293729
#> [21,] 0.4800702 0.6128094 0.7112174  0.81473228  0.83612757 0.92052395
#> [22,] 0.3998342 0.4038686 0.5132973  0.69114030  0.45319242 0.60251086
#> [23,] 0.2415619 0.2400691 0.3488488  0.41787343  0.53470238 0.63674662
#> [24,] 0.3501269 0.4831784 0.6617708  0.78151206  0.79786967 0.79711210
#> [25,] 0.3993758 0.4129940 0.4344162  0.60152003  0.73732166 0.85868636
#>             V26        V27        V28         V29            V3       V30
#>  [1,] 0.9134798 0.91765189 0.88953667  0.79316269  0.0333712435 0.6001774
#>  [2,] 0.9451380 0.96100695 0.88231073  0.73528144  0.0175733256 0.4868428
#>  [3,] 0.9749295 0.99903251 0.97330927  0.87783430  0.0323834325 0.6334210
#>  [4,] 0.8201316 0.91073194 0.92770546  0.90230070  0.0167122389 0.9463104
#>  [5,] 0.5006595 0.46121239 0.39331059  0.43720585  0.0781262811 0.3517088
#>  [6,] 0.6118758 0.30290470 0.23286200  0.31278183  0.0648994032 0.3310228
#>  [7,] 0.1495102 0.09230354 0.02122252  0.09092844  0.0647507771 0.2653575
#>  [8,] 0.2468339 0.17605949 0.15856002 -0.07188972  0.1721896336 0.2541847
#>  [9,] 0.8608593 0.72501703 0.64021929  0.56663115  0.0246605052 0.5011060
#> [10,] 0.8636389 0.92753152 0.93637291  0.85173415  0.0406855644 0.7065645
#> [11,] 0.6309408 0.69949081 0.79897509  0.78379147  0.0833659802 0.8043229
#> [12,] 0.5269557 0.54915008 0.70495335  0.66436947  0.0323504409 0.6095851
#> [13,] 0.5965283 0.50084863 0.45007182  0.45850750  0.0762434220 0.4028560
#> [14,] 0.5240081 0.40551637 0.32633471  0.36793740  0.0759372021 0.3769185
#> [15,] 0.3014311 0.34694890 0.50901061  0.60243080  0.0303363805 0.6190440
#> [16,] 0.2580189 0.46422971 0.77014507  0.92931500  0.0782473979 0.7910223
#> [17,] 0.6131696 0.61225784 0.67769084  0.77525801  0.0573336821 0.8332450
#> [18,] 0.7526662 0.78068184 0.59081563  0.26551196  0.0769655373 0.2975648
#> [19,] 0.1627648 0.40499631 0.64148285  0.79559998 -0.0085703532 0.7307765
#> [20,] 0.9044618 0.92496709 0.91104618  0.82800034  0.0005105402 0.7086404
#> [21,] 0.9736686 0.95738955 0.82785604  0.72371643  0.0270161197 0.6467704
#> [22,] 0.9039425 0.84694852 0.76351387  0.48205243  0.0140389234 0.4444518
#> [23,] 0.6542115 0.59699195 0.68545854  0.82573568  0.0189145008 0.8831584
#> [24,] 0.8563132 0.83171647 0.80856024  0.81901634  0.0203900854 0.7582073
#> [25,] 0.8820670 0.84425758 0.85572001  0.84443382  0.0076610657 0.6952326
#>             V31       V32       V33       V34        V35        V36        V37
#>  [1,] 0.4258648 0.3860191 0.3130848 0.1611244 0.06683302 0.18747540 0.21977494
#>  [2,] 0.1198622 0.2824482 0.4607665 0.6014886 0.81745513 1.00579299 0.89638352
#>  [3,] 0.2310071 0.2656700 0.3419422 0.4973086 0.76843869 0.90840205 0.80376289
#>  [4,] 0.9224407 0.8396020 0.7274262 0.6628048 0.67022094 0.71740592 0.69033729
#>  [5,] 0.2245124 0.2525870 0.3621466 0.4008198 0.47983060 0.49617868 0.48397796
#>  [6,] 0.3985113 0.4537413 0.5373182 0.5356724 0.52357671 0.38890628 0.26198115
#>  [7,] 0.2662376 0.2929509 0.2952201 0.4045898 0.37280458 0.31428045 0.46313736
#>  [8,] 0.1481083 0.1750970 0.2788943 0.2313571 0.36360829 0.06049685 0.01937283
#>  [9,] 0.3669460 0.2010612 0.1370776 0.2290194 0.41718146 0.49236543 0.36559325
#> [10,] 0.5973463 0.4724470 0.3920225 0.2911059 0.18015395 0.09587541 0.16385653
#> [11,] 0.7950153 0.7201495 0.4967537 0.4044889 0.36105665 0.22046275 0.24482931
#> [12,] 0.3328117 0.2367976 0.3893950 0.5868190 0.74535243 0.77221575 0.79276642
#> [13,] 0.4090673 0.1600896 0.4621208 0.6599676 0.60731648 0.44790129 0.42431679
#> [14,] 0.2895875 0.2034657 0.2940823 0.4193664 0.43058205 0.39848135 0.35027587
#> [15,] 0.6530848 0.6164981 0.6452069 0.6700189 0.72330406 0.74765633 0.75044574
#> [16,] 0.4946263 0.2471384 0.5303527 0.6231711 0.34358978 0.44098272 0.62116332
#> [17,] 0.6682665 0.3820446 0.3035401 0.3746016 0.31673691 0.22926504 0.36288527
#> [18,] 0.5759594 0.5606835 0.2629850 0.3595447 0.75247684 0.70527743 0.51357188
#> [19,] 0.5159039 0.1526995 0.4587738 0.4573243 0.22136069 0.34874544 0.48209525
#> [20,] 0.5906186 0.5059961 0.4876274 0.5513782 0.59961179 0.54387572 0.52021123
#> [21,] 0.5851283 0.5301148 0.5277943 0.5373267 0.55661108 0.53566530 0.46595854
#> [22,] 0.5571448 0.5701130 0.3941695 0.3549869 0.73370039 0.77536673 0.40177142
#> [23,] 0.8103583 0.7603544 0.7858495 0.7415090 0.71635958 0.63707137 0.48519994
#> [24,] 0.6777599 0.6189781 0.6320002 0.6852070 0.69829433 0.62609752 0.49943586
#> [25,] 0.5348831 0.3827643 0.3872261 0.2838741 0.23415956 0.25738929 0.19503069
#>              V38       V39          V4          V40        V41        V42
#>  [1,] 0.27127604 0.3849659 0.054775660  0.361214913 0.24663705 0.29300169
#>  [2,] 0.65088058 0.4281499 0.036879279  0.256343477 0.24508379 0.32038545
#>  [3,] 0.62774351 0.4562223 0.029712961  0.311999665 0.28724265 0.24644332
#>  [4,] 0.56815623 0.4878453 0.053195691  0.499724400 0.47762058 0.42178767
#>  [5,] 0.55122772 0.4589594 0.114974987  0.449405024 0.43310862 0.47081022
#>  [6,] 0.42807650 0.4430633 0.048816520  0.458366667 0.38371699 0.35092418
#>  [7,] 0.44698828 0.1819422 0.051614371  0.341664249 0.63328885 0.45597235
#>  [8,] 0.24950805 0.2058180 0.147849461 -0.006336789 0.11063675 0.19405098
#>  [9,] 0.26237837 0.3857841 0.095973110  0.442643952 0.48424386 0.57767319
#> [10,] 0.17092355 0.1885158 0.048006622  0.188875295 0.21143355 0.21920623
#> [11,] 0.35470199 0.3659141 0.078164484  0.239178376 0.52570481 0.64125537
#> [12,] 0.80267364 0.7170544 0.050965870  0.512280818 0.27226937 0.20480671
#> [13,] 0.35921253 0.4368436 0.078559862  0.414130300 0.26934170 0.36773305
#> [14,] 0.33540416 0.3571748 0.105693813  0.256509696 0.14513551 0.24670032
#> [15,] 0.79758812 0.8521898 0.017012704  0.836491118 0.70268525 0.61587039
#> [16,] 0.62956157 0.5416711 0.079482607  0.321838013 0.23719124 0.06812049
#> [17,] 0.32575467 0.2251099 0.040590895  0.270302969 0.22066874 0.19078208
#> [18,] 0.53192425 0.4116998 0.065271658  0.291176082 0.12926990 0.16686777
#> [19,] 0.58725786 0.6741070 0.005196195  0.530746124 0.20897622 0.20506005
#> [20,] 0.36207631 0.2049708 0.028257789  0.090196873 0.09575215 0.03130524
#> [21,] 0.33287560 0.1443379 0.023359742  0.193888899 0.22292736 0.17556232
#> [22,] 0.06889334 0.3807637 0.028141486  0.626859150 0.35796715 0.22341888
#> [23,] 0.35705365 0.2721808 0.023164154  0.253800503 0.23657795 0.20238292
#> [24,] 0.39441983 0.2980051 0.025328349  0.278856224 0.26468675 0.20145328
#> [25,] 0.16053171 0.2137062 0.008840570  0.151741858 0.20368760 0.25942811
#>                V43         V44          V45         V46          V47
#>  [1,]  0.315198126  0.31881032  0.291358947  0.17069147  0.127273499
#>  [2,]  0.425306139  0.45610994  0.399139860  0.25252938  0.106729326
#>  [3,]  0.323180411  0.36852827  0.295399354  0.21054420  0.081431903
#>  [4,]  0.451045180  0.52758611  0.473825793  0.31504299  0.210454694
#>  [5,]  0.485493182  0.32295257  0.149308403  0.11646028  0.117403940
#>  [6,]  0.241112968  0.18062159  0.194899499  0.20847809  0.121699914
#>  [7,]  0.187635479  0.14139875  0.058807499  0.05867821  0.067610472
#>  [8,]  0.144880509  0.14444433  0.067891636  0.01869989  0.094824075
#>  [9,]  0.438858065  0.24598868  0.105306282  0.20162459  0.233332824
#> [10,]  0.207704723  0.16291339  0.132129822  0.14228028  0.133887519
#> [11,]  0.524459772  0.43842671  0.566868440  0.54993948  0.354864492
#> [12,]  0.254978374  0.33995041  0.435670709  0.34762315  0.204671990
#> [13,]  0.362576034  0.18175598  0.142996106  0.02478645  0.110248066
#> [14,]  0.332313046  0.21401617  0.098874149  0.08560929  0.088855892
#> [15,]  0.453643398  0.30836886  0.339662289  0.31236027  0.254292415
#> [16,] -0.003427294 -0.01352697 -0.026240112 -0.02012798 -0.008077301
#> [17,]  0.154552574  0.11210010  0.183609632  0.16597634  0.159731461
#> [18,]  0.262292437  0.26400442  0.208557772  0.08169516  0.069253577
#> [19,]  0.162028297  0.05170371 -0.009141903 -0.02666716  0.062762878
#> [20,]  0.092683392  0.04283994 -0.005017503  0.03444011  0.034706826
#> [21,]  0.114739013  0.07624540  0.079748950  0.05737152  0.030746621
#> [22,]  0.217842605  0.30141681  0.181198910  0.14092983  0.040523998
#> [23,]  0.188554044  0.22088264  0.194230726  0.12371804  0.077226205
#> [24,]  0.232627971  0.17083573  0.114162346  0.08838709  0.060304966
#> [25,]  0.219542996  0.12620221  0.077307648  0.08969194  0.065971122
#>                V48          V49         V5          V50           V51
#>  [1,]  0.116833281  0.067517051 0.09101925  0.018938492  2.241394e-02
#>  [2,]  0.034614233  0.028469872 0.04551238  0.009186395  4.516047e-03
#>  [3,]  0.034736790  0.029047474 0.04939321  0.012075452  4.014281e-03
#>  [4,]  0.181036397  0.101478550 0.08994604  0.027538592  1.575428e-02
#>  [5,]  0.065534785  0.040019288 0.14720380  0.022275706  1.582057e-02
#>  [6,]  0.073519137  0.055577841 0.11809524  0.010485812  8.449022e-05
#>  [7,]  0.025532724  0.024025934 0.05944655  0.019372893  6.120780e-03
#>  [8,]  0.084466090  0.014120802 0.15288444  0.010180720  2.964327e-02
#>  [9,]  0.133564405  0.068719146 0.12919595  0.038138011  2.761545e-02
#> [10,]  0.102754530  0.052016112 0.07224450  0.015869552  1.736604e-02
#> [11,]  0.219689469  0.130588860 0.11242697  0.049406735  4.523282e-02
#> [12,]  0.151781179  0.086010298 0.05169433  0.035138081  1.843879e-02
#> [13,]  0.187023125  0.095480020 0.02218968  0.020127792  1.888440e-02
#> [14,]  0.050456263  0.055479694 0.14426889  0.021376700  1.850000e-02
#> [15,]  0.169239458  0.102437513 0.04503083  0.039559878  2.426369e-02
#> [16,] -0.003306471 -0.002731734 0.05031154 -0.001668757 -1.027108e-03
#> [17,]  0.104342362  0.050606910 0.03588538  0.029378397  1.059670e-02
#> [18,]  0.061278872  0.033028935 0.10838393  0.025417206  6.908488e-03
#> [19,]  0.073155912  0.016226560 0.04008894  0.011429118  6.033667e-03
#> [20,]  0.032727957  0.018557789 0.05866884  0.008491506  9.336315e-04
#> [21,]  0.034571847  0.012921670 0.05569482  0.004488956  2.154012e-03
#> [22,]  0.048976416  0.012845304 0.10228274  0.006368116  1.882517e-02
#> [23,]  0.051601631  0.027150825 0.03282727  0.009931283  9.466562e-03
#> [24,]  0.045803169  0.019559479 0.04735947  0.014991031  7.204307e-03
#> [25,]  0.040099220  0.020786144 0.02704232  0.016891239  1.052049e-02
#>               V52         V53         V54         V55          V56          V57
#>  [1,] 0.013572010 0.009456639 0.012807231 0.008316114 0.0081943167  0.005482778
#>  [2,] 0.009761000 0.004812593 0.004796547 0.001641977 0.0005271512  0.003973326
#>  [3,] 0.010224647 0.008091751 0.007199819 0.003576162 0.0035947833  0.001258804
#>  [4,] 0.013826477 0.011716089 0.008508182 0.004769560 0.0073332454  0.006507267
#>  [5,] 0.009122581 0.018939501 0.025633367 0.004647730 0.0151849738  0.008033668
#>  [6,] 0.014758317 0.016991520 0.002678193 0.006258949 0.0106322796  0.004294567
#>  [7,] 0.032785083 0.019189065 0.013957268 0.028228988 0.0150872648  0.004945623
#>  [8,] 0.013922044 0.029448639 0.012574209 0.015794542 0.0172548187  0.022420741
#>  [9,] 0.024227674 0.030025091 0.018899438 0.001598429 0.0065545767  0.013668423
#> [10,] 0.013174542 0.008544947 0.007143832 0.005776795 0.0057435862  0.004996588
#> [11,] 0.030008955 0.019573896 0.017885186 0.015643757 0.0129695318  0.014061351
#> [12,] 0.012410555 0.010556511 0.014094093 0.012540620 0.0110123903  0.010080797
#> [13,] 0.017038191 0.017438570 0.021558767 0.014193845 0.0065033086  0.003577543
#> [14,] 0.011628337 0.007684212 0.026171434 0.017138816 0.0094387808  0.016103559
#> [15,] 0.010484841 0.010800697 0.011611610 0.013198011 0.0108475529  0.004278392
#> [16,] 0.004948680 0.024525360 0.019914400 0.003681065 0.0139082068  0.016634429
#> [17,] 0.008070945 0.006649580 0.012353618 0.015821343 0.0121912651  0.013311306
#> [18,] 0.018807248 0.005488934 0.019318924 0.012692971 0.0031056154  0.004693635
#> [19,] 0.022550971 0.029056142 0.016272683 0.025475201 0.0025151009 -0.000661197
#> [20,] 0.003433717 0.008822920 0.009651771 0.009396372 0.0022276995  0.003402390
#> [21,] 0.003633320 0.008532303 0.003443325 0.005782530 0.0053753429  0.002553798
#> [22,] 0.004968226 0.010026634 0.019170928 0.007208386 0.0161611864  0.015478972
#> [23,] 0.006560325 0.010328495 0.007972415 0.007246922 0.0062189430  0.005867129
#> [24,] 0.006609395 0.008164821 0.008471634 0.006290817 0.0070465856  0.009791278
#> [25,] 0.007716389 0.006558326 0.004600240 0.005998454 0.0059905913  0.006205978
#>               V58         V59         V6         V60          V7           V8
#>  [1,] 0.007967604 0.006182924 0.15302560 0.006020637  0.15811791  0.197056918
#>  [2,] 0.002274070 0.002456535 0.07444617 0.004020244  0.07982790  0.043186337
#>  [3,] 0.000979162 0.003225158 0.07047612 0.004197721  0.08560945  0.020076570
#>  [4,] 0.004469475 0.001591900 0.11375017 0.002889789  0.15194280  0.158382770
#>  [5,] 0.006609321 0.017093716 0.09826496 0.010031935  0.01744017  0.111323027
#>  [6,] 0.007735869 0.013533989 0.10809025 0.004449270  0.13516893  0.031357107
#>  [7,] 0.012308127 0.012608127 0.05190136 0.005531797  0.06219005  0.058257048
#>  [8,] 0.006870059 0.015810923 0.05593827 0.016112680  0.04126510 -0.007277403
#>  [9,] 0.010634874 0.011407987 0.05200702 0.009339873  0.12995068  0.119189949
#> [10,] 0.005150194 0.005051317 0.08806264 0.004834286  0.09930235  0.114976624
#> [11,] 0.016053051 0.013076809 0.10292906 0.009646883  0.10545561  0.138769745
#> [12,] 0.008943992 0.007563137 0.06236436 0.006370588  0.10802951  0.199874496
#> [13,] 0.016654186 0.008113049 0.03111332 0.003354059  0.08383328  0.240219156
#> [14,] 0.029127197 0.013871466 0.14768950 0.004970773  0.17230309  0.162227046
#> [15,] 0.009624997 0.008140265 0.08425658 0.008412087  0.13355988  0.147161827
#> [16,] 0.007631374 0.013866207 0.10290820 0.011078468 -0.00144311  0.041417653
#> [17,] 0.007310294 0.005716581 0.11689506 0.009424887  0.15077667  0.166060435
#> [18,] 0.005023986 0.011963424 0.23046773 0.008131514  0.22476549  0.207754323
#> [19,] 0.003874586 0.008444481 0.04656507 0.009045672  0.02815458  0.058572749
#> [20,] 0.001593896 0.002185488 0.10051255 0.002906018  0.10853530  0.087986726
#> [21,] 0.004080962 0.007071704 0.09546374 0.007219427  0.09923431  0.074755695
#> [22,] 0.005415781 0.017220254 0.15785843 0.008520856  0.12775237  0.109939888
#> [23,] 0.005792844 0.006726085 0.04569537 0.003159405  0.06728625  0.037077969
#> [24,] 0.006729785 0.005022073 0.08034312 0.006170436  0.09515871  0.095193828
#> [25,] 0.004243747 0.002995747 0.06548041 0.002707252  0.08969694  0.061249291
#>                V9
#>  [1,] 0.276544285
#>  [2,] 0.034176453
#>  [3,] 0.120294227
#>  [4,] 0.289461702
#>  [5,] 0.013439548
#>  [6,] 0.056151764
#>  [7,] 0.140574256
#>  [8,] 0.151424685
#>  [9,] 0.015878371
#> [10,] 0.173344643
#> [11,] 0.206097443
#> [12,] 0.310686106
#> [13,] 0.430928022
#> [14,] 0.200822986
#> [15,] 0.009551598
#> [16,] 0.098512238
#> [17,] 0.272325667
#> [18,] 0.323795010
#> [19,] 0.143140203
#> [20,] 0.157360093
#> [21,] 0.109813632
#> [22,] 0.098367333
#> [23,] 0.049328406
#> [24,] 0.096158132
#> [25,] 0.090331991
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058