Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

LearnerClassifLvq1$new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>               V1        V10        V11        V12        V13       V14
#>  [1,] 0.04018070 0.41413376 0.51568515 0.55456750 0.52047624 0.4091370
#>  [2,] 0.09220340 0.57973506 0.47310984 0.20841763 0.21059019 0.3121174
#>  [3,] 0.03858103 0.29060058 0.32118496 0.34767882 0.31636446 0.3595402
#>  [4,] 0.02146435 0.27207413 0.31159165 0.33333187 0.27767754 0.1950336
#>  [5,] 0.01331515 0.20021600 0.18842898 0.16673763 0.37462902 0.5828183
#>  [6,] 0.02971441 0.22757805 0.27616774 0.32895258 0.34922882 0.3862785
#>  [7,] 0.04980668 0.08436889 0.18640217 0.27438361 0.27501469 0.3396125
#>  [8,] 0.03259139 0.26188728 0.33430647 0.35044559 0.27508966 0.1485588
#>  [9,] 0.01717119 0.30540896 0.33682817 0.30653687 0.30469822 0.2578459
#> [10,] 0.01572003 0.12816724 0.19833978 0.24585104 0.14913114 0.1119573
#> [11,] 0.05158683 0.22903447 0.30336943 0.28619270 0.28567946 0.2928508
#> [12,] 0.02493060 0.27438146 0.28337109 0.29320186 0.27969125 0.2195628
#> [13,] 0.04291962 0.71685680 0.74227758 0.50727056 0.30130669 0.1961095
#> [14,] 0.05337246 0.20028668 0.34777936 0.49271657 0.59633567 0.6174396
#> [15,] 0.02068029 0.07887799 0.17994280 0.33238904 0.24876081 0.1876684
#> [16,] 0.02725067 0.10296194 0.09164897 0.16119245 0.16420230 0.1141893
#> [17,] 0.00990000 0.18820000 0.14560000 0.18920000 0.31760000 0.1340000
#> [18,] 0.01260000 0.42840000 0.30150000 0.12070000 0.32990000 0.5707000
#> [19,] 0.01881754 0.09771239 0.08789032 0.06500016 0.07559512 0.1051414
#> [20,] 0.03987691 0.24522657 0.26704690 0.32066814 0.27086348 0.2617261
#> [21,] 0.02368211 0.25701820 0.23155884 0.11512266 0.18336051 0.4775250
#> [22,] 0.01191060 0.10387735 0.10634324 0.07983666 0.08413177 0.1169829
#> [23,] 0.01034371 0.10343204 0.09724374 0.07045885 0.09619401 0.1176151
#> [24,] 0.02186298 0.12215157 0.15473839 0.18410483 0.19216260 0.1855166
#> [25,] 0.02566929 0.20716748 0.28237989 0.22942986 0.26732503 0.4046389
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.26559533 0.2628662 0.3113840 0.3177501 0.4381220 0.041934619 0.6371801
#>  [2,] 0.54407790 0.5570654 0.1882911 0.1415203 0.3794449 0.097829901 0.8807986
#>  [3,] 0.30819226 0.2142779 0.2029944 0.1975814 0.3345271 0.042457046 0.4541677
#>  [4,] 0.16138107 0.1947499 0.2427212 0.3467203 0.5007739 0.034736530 0.6126924
#>  [5,] 0.60446630 0.6130042 0.6722998 0.7846929 0.8524364 0.028574509 0.9041966
#>  [6,] 0.44332962 0.5037910 0.5890509 0.6635529 0.6880070 0.025762413 0.7644947
#>  [7,] 0.50052958 0.6407457 0.8254070 0.8287945 0.8673711 0.056471315 0.9228784
#>  [8,] 0.14352035 0.1711769 0.1979117 0.2464693 0.2303744 0.033668086 0.2545450
#>  [9,] 0.15842508 0.1171657 0.2639009 0.4304395 0.6487701 0.044102640 0.7618593
#> [10,] 0.19261246 0.3259978 0.4142856 0.3510930 0.4239868 0.049475984 0.3880514
#> [11,] 0.26509603 0.2263872 0.2543927 0.2043819 0.2911057 0.044148943 0.3204163
#> [12,] 0.18194351 0.2251313 0.2495813 0.2139435 0.1871204 0.036589058 0.1765017
#> [13,] 0.27891521 0.3761433 0.2470229 0.2442376 0.3227469 0.090934665 0.4813087
#> [14,] 0.67962035 0.7771241 0.9162171 0.9524212 0.9363843 0.066137347 0.8399821
#> [15,] 0.10148182 0.1338023 0.1970247 0.4318831 0.3527315 0.053735752 0.5913172
#> [16,] 0.16376455 0.1750586 0.1641357 0.1192317 0.1715679 0.042159358 0.2314431
#> [17,] 0.21690000 0.2458000 0.2589000 0.2786000 0.2298000 0.048400000 0.0656000
#> [18,] 0.69620000 0.9751000 1.0000000 0.9293000 0.6210000 0.051900000 0.4586000
#> [19,] 0.04652566 0.1409085 0.1662570 0.3657130 0.3458877 0.019111051 0.3217037
#> [20,] 0.34257742 0.4316200 0.4847893 0.5982521 0.7370211 0.058345283 0.8466034
#> [21,] 0.67412438 0.7237722 0.7487488 0.6001455 0.3942586 0.061031553 0.3188874
#> [22,] 0.10686011 0.1268844 0.1881325 0.2427481 0.2783893 0.009543066 0.3493512
#> [23,] 0.16338695 0.1807863 0.2049983 0.2568843 0.3064471 0.009872459 0.3174253
#> [24,] 0.23442009 0.4564046 0.5914176 0.4744727 0.3277769 0.019961628 0.3548336
#> [25,] 0.51154841 0.6648692 0.8279798 0.8979109 0.8235086 0.033095899 0.6283049
#>             V21       V22       V23       V24        V25       V26       V27
#>  [1,] 0.8011006 0.8458131 0.7620496 0.7643962 0.75349399 0.8332342 0.9061472
#>  [2,] 0.8898323 0.3891664 0.1495351 0.1509388 0.21285495 0.3632644 0.2625684
#>  [3,] 0.6984782 0.7772997 0.7140634 0.7295492 0.74803655 0.6505744 0.7428435
#>  [4,] 0.7591486 0.8038299 0.8416503 0.9672555 0.98635529 0.9755342 0.9735619
#>  [5,] 0.9991129 1.0462222 0.9785129 0.8644963 0.80807437 0.7701249 0.6802689
#>  [6,] 0.8581987 0.9104050 0.9398860 0.9935368 0.99539618 0.9673549 0.9278587
#>  [7,] 0.9481340 0.9536893 0.9019889 0.8601731 0.73298508 0.6055334 0.4830689
#>  [8,] 0.4022617 0.5468485 0.6679569 0.7612826 0.84288141 0.8825252 0.9356350
#>  [9,] 0.6914953 0.4483853 0.2323432 0.2686263 0.50876295 0.8468727 1.0243076
#> [10,] 0.5810154 0.6161402 0.6359888 0.6871278 0.61095758 0.6439866 0.8108993
#> [11,] 0.3301521 0.3342257 0.3927529 0.3934549 0.45462816 0.5843787 0.7318999
#> [12,] 0.2212024 0.3523190 0.4948666 0.6134784 0.67575591 0.7627248 0.8540997
#> [13,] 0.6493501 0.8465983 0.9759806 1.0049183 0.67125383 0.5373978 0.4035814
#> [14,] 0.7244033 0.6195321 0.5253263 0.4374789 0.30425293 0.2119559 0.1928781
#> [15,] 0.7271246 0.7700223 0.8161005 0.8368889 0.92956260 0.8258010 0.6734755
#> [16,] 0.2072940 0.2131877 0.2882520 0.2396248 0.19971221 0.3535357 0.5330570
#> [17,] 0.1441000 0.1179000 0.1668000 0.1783000 0.24760000 0.2570000 0.1036000
#> [18,] 0.5001000 0.5032000 0.7082000 0.8420000 0.81090000 0.7690000 0.8105000
#> [19,] 0.2348926 0.1862893 0.1787330 0.2333213 0.07154674 0.1996214 0.4181738
#> [20,] 0.9018045 0.8327281 0.7481597 0.7262724 0.79106697 0.8066158 0.6972411
#> [21,] 0.5444298 0.7789866 0.8639509 0.9322630 0.79798931 0.6882911 0.6887676
#> [22,] 0.4395644 0.5927090 0.7159260 0.7596413 0.78543589 0.8529676 0.8534829
#> [23,] 0.3426532 0.3873639 0.4768830 0.6632856 0.84695490 0.8996831 0.8575892
#> [24,] 0.4011739 0.4734518 0.5396949 0.6224628 0.75411156 0.8251490 0.8215104
#> [25,] 0.6298917 0.6666345 0.6766392 0.7167491 0.66689143 0.5952759 0.5668072
#>             V28       V29          V3       V30       V31       V32       V33
#>  [1,] 0.8748237 0.7308191 0.050937369 0.5362374 0.4165098 0.4299824 0.4002538
#>  [2,] 0.5676621 0.5500085 0.052542311 0.7197423 0.3800392 0.4137806 0.6543633
#>  [3,] 0.7664488 0.9575409 0.047327243 1.0447850 0.8788169 0.5907233 0.4893664
#>  [4,] 0.9454684 0.8464411 0.034554181 0.6694540 0.4635660 0.3254968 0.2226031
#>  [5,] 0.6664771 0.6598884 0.020939471 0.4592364 0.2525310 0.1774104 0.1493847
#>  [6,] 0.9090080 0.7995428 0.023673271 0.6281203 0.4771438 0.3625759 0.2553502
#>  [7,] 0.3458052 0.3076285 0.072177762 0.3027410 0.2643275 0.3651523 0.4396851
#>  [8,] 0.9706801 0.8724251 0.054268402 0.6735837 0.6520626 0.5017120 0.3504499
#>  [9,] 0.9313226 0.8660088 0.046233410 0.7894279 0.5768330 0.5096996 0.6994824
#> [10,] 0.7849480 0.6733380 0.056400682 0.8092875 0.6305024 0.5427733 0.3611833
#> [11,] 0.9014297 0.9233897 0.028409115 0.7795423 0.6267923 0.5574614 0.5150544
#> [12,] 0.8996260 0.8896807 0.041240868 0.9151652 0.9162288 0.8394982 0.7004023
#> [13,] 0.3437569 0.4234345 0.083326171 0.4142762 0.4335872 0.1046734 0.4273039
#> [14,] 0.2029105 0.2066861 0.069992230 0.1946121 0.2668770 0.3384654 0.3344547
#> [15,] 0.5759057 0.5171542 0.049887212 0.5017159 0.5648870 0.5430136 0.5928006
#> [16,] 0.6597278 0.7208275 0.024773101 0.7945207 0.7759906 0.6228532 0.7186614
#> [17,] 0.5356000 0.7124000 0.029900000 0.6291000 0.4756000 0.6015000 0.7208000
#> [18,] 0.6203000 0.2356000 0.062100000 0.2595000 0.6299000 0.6762000 0.2903000
#> [19,] 0.6611429 0.8106944 0.014448924 0.7395400 0.5141191 0.1980397 0.4780842
#> [20,] 0.5365153 0.3134990 0.040612597 0.1408847 0.3137233 0.6922830 0.7494234
#> [21,] 0.5712168 0.4292313 0.106867303 0.4701369 0.4550124 0.2239920 0.1854056
#> [22,] 0.8534217 0.8667179 0.015490068 0.7902270 0.6841399 0.6263972 0.6344371
#> [23,] 0.7201419 0.5377006 0.006755691 0.4035719 0.3285425 0.2538314 0.2383513
#> [24,] 0.8412962 0.7732051 0.027452189 0.7369497 0.7778964 0.7223210 0.6375471
#> [25,] 0.4983889 0.4454538 0.064643106 0.5220673 0.4001922 0.3024021 0.2321424
#>              V34        V35        V36       V37        V38       V39
#>  [1,] 0.27236520 0.07333921 0.22063441 0.2586540 0.30626683 0.4938230
#>  [2,] 0.78952572 0.89428822 0.86224469 0.8571554 0.89702818 0.7286204
#>  [3,] 0.46827005 0.29535698 0.30866845 0.2287769 0.06897898 0.1667720
#>  [4,] 0.19834545 0.17021969 0.08775618 0.1053940 0.14155789 0.1462700
#>  [5,] 0.07888929 0.11938230 0.22521733 0.1816755 0.20605159 0.2917867
#>  [6,] 0.09127027 0.06678492 0.10790458 0.2044815 0.27222242 0.2581256
#>  [7,] 0.42867595 0.41586852 0.32672755 0.3348233 0.33122513 0.3745842
#>  [8,] 0.20889632 0.09688936 0.06643055 0.2065736 0.16212598 0.1723768
#>  [9,] 0.65338097 0.52164092 0.35085317 0.3709479 0.38549467 0.3436187
#> [10,] 0.40347951 0.28619805 0.08630446 0.2660003 0.40852892 0.2415660
#> [11,] 0.44182530 0.32810662 0.23033834 0.2108040 0.23327527 0.1923286
#> [12,] 0.64864060 0.67916688 0.63992493 0.5438147 0.48183087 0.4047477
#> [13,] 0.65283803 0.61743180 0.43452024 0.4075621 0.35729921 0.4539959
#> [14,] 0.39048720 0.41148681 0.23970765 0.2498917 0.32968881 0.2835720
#> [15,] 0.53325951 0.58415808 0.60187053 0.4718104 0.31431039 0.1347098
#> [16,] 0.89150212 0.95798971 0.92973611 0.7035583 0.54812135 0.4279089
#> [17,] 0.62340000 0.57250000 0.75230000 0.8712000 0.92520000 0.9709000
#> [18,] 0.43930000 0.85290000 0.71800000 0.4801000 0.58560000 0.4993000
#> [19,] 0.48982227 0.25433423 0.36092473 0.4963029 0.58282398 0.6332216
#> [20,] 0.78019926 0.74451205 0.72611067 0.6181800 0.22255061 0.3573962
#> [21,] 0.08312968 0.38546865 0.67376768 0.6986957 0.34539054 0.1524843
#> [22,] 0.65982625 0.62709954 0.54415289 0.4090565 0.32903937 0.2456747
#> [23,] 0.25196037 0.26571691 0.26045284 0.2560215 0.26004908 0.2838347
#> [24,] 0.64174281 0.70435378 0.70600234 0.6211954 0.46011053 0.3336153
#> [25,] 0.16414456 0.32978357 0.43801008 0.4222009 0.27642550 0.1746715
#>                V4       V40        V41        V42       V43        V44
#>  [1,] 0.039213297 0.4191451 0.20230562 0.33112640 0.4285640 0.46547127
#>  [2,] 0.101921541 0.4976069 0.30960075 0.29096517 0.2536156 0.30113312
#>  [3,] 0.057156435 0.2982198 0.37085277 0.39511719 0.3365895 0.07220555
#>  [4,] 0.026886092 0.1433913 0.05368088 0.09863148 0.1950054 0.21984130
#>  [5,] 0.045866326 0.2714813 0.25736273 0.25224039 0.1844748 0.20945452
#>  [6,] 0.055832583 0.2977817 0.27181146 0.26723067 0.2550552 0.19903828
#>  [7,] 0.051679291 0.3706758 0.31486270 0.34835268 0.2385453 0.26637758
#>  [8,] 0.060765362 0.1703600 0.19208951 0.23359860 0.1670965 0.09850943
#>  [9,] 0.068958535 0.2829396 0.37725748 0.39292817 0.3607248 0.21834998
#> [10,] 0.047612854 0.3874092 0.41774113 0.19333072 0.3033770 0.33882698
#> [11,] 0.060460816 0.2507124 0.42867527 0.56216989 0.5153288 0.31479042
#> [12,] 0.060958062 0.4235855 0.51149669 0.48790323 0.4271379 0.49529392
#> [13,] 0.080479344 0.4349260 0.29433434 0.43591270 0.4202295 0.18108583
#> [14,] 0.085428059 0.1853674 0.15499455 0.21112623 0.1860793 0.07522964
#> [15,] 0.054374157 0.3265195 0.37111606 0.48791009 0.4697816 0.28306130
#> [16,] 0.011979468 0.5314222 0.50288153 0.46790802 0.3083929 0.21196666
#> [17,] 0.029700000 0.9297000 0.89950000 0.79110000 0.5600000 0.28380000
#> [18,] 0.051800000 0.2866000 0.06010000 0.11670000 0.2737000 0.28120000
#> [19,] 0.025682063 0.4839926 0.22601333 0.18344368 0.1342315 0.05457481
#> [20,] 0.072957326 0.5265662 0.49044121 0.41817685 0.3460085 0.35583456
#> [21,] 0.094772372 0.3568300 0.43941788 0.36678326 0.2353420 0.18857841
#> [22,] 0.025817913 0.2295734 0.21353892 0.15808275 0.1582709 0.11897228
#> [23,] 0.004183581 0.2723666 0.24789347 0.20769776 0.1559228 0.12342066
#> [24,] 0.036949580 0.2708054 0.23192917 0.22265640 0.2175942 0.21688282
#> [25,] 0.064632853 0.3006887 0.34565813 0.26617668 0.1509909 0.16352645
#>              V45         V46        V47        V48        V49         V5
#>  [1,] 0.43036634 0.203827340 0.10900860 0.12689877 0.08105705 0.08162969
#>  [2,] 0.44104387 0.324235313 0.13551557 0.10289996 0.05555396 0.08539399
#>  [3,] 0.13470070 0.054033528 0.08667651 0.11087540 0.07017691 0.07269774
#>  [4,] 0.18693115 0.200191216 0.15939384 0.08555268 0.04566276 0.05538565
#>  [5,] 0.16084518 0.127736471 0.13594666 0.11232359 0.08425439 0.12080935
#>  [6,] 0.13979236 0.108565772 0.12505096 0.09887143 0.05858239 0.11207320
#>  [7,] 0.21251555 0.110463496 0.06781447 0.03425765 0.03509881 0.10461063
#>  [8,] 0.08489757 0.176609172 0.21132480 0.16051308 0.07112305 0.07442484
#>  [9,] 0.10573195 0.079920209 0.06972039 0.05058175 0.02187929 0.17121798
#> [10,] 0.21102576 0.132618470 0.08515788 0.10134121 0.08338183 0.02651565
#> [11,] 0.34794170 0.379664678 0.30922628 0.22793268 0.12626450 0.06653623
#> [12,] 0.49906127 0.377248593 0.23660942 0.16780776 0.08362659 0.07937253
#> [13,] 0.14792234 0.010986588 0.11744685 0.20864526 0.10808332 0.01347535
#> [14,] 0.07415727 0.075586550 0.05921935 0.05437363 0.04197608 0.10158895
#> [15,] 0.18005850 0.146841515 0.09528717 0.07029013 0.03080310 0.10629904
#> [16,] 0.18156074 0.137790055 0.12904232 0.08384287 0.04515775 0.03136134
#> [17,] 0.44070000 0.550700000 0.43310000 0.29050000 0.19810000 0.06520000
#> [18,] 0.20780000 0.066000000 0.04910000 0.03450000 0.01720000 0.10720000
#> [19,] 0.01538513 0.001645292 0.05979065 0.06438897 0.01818365 0.04375883
#> [20,] 0.16085558 0.067057443 0.07108208 0.04338235 0.04508509 0.08341750
#> [21,] 0.20790712 0.167430140 0.17521239 0.15883319 0.11308680 0.05762700
#> [22,] 0.07256057 0.042871852 0.02981012 0.03231979 0.01674286 0.04359350
#> [23,] 0.09670626 0.078140562 0.02466372 0.02125011 0.01362606 0.02169011
#> [24,] 0.17248280 0.152552274 0.08284905 0.05531217 0.03247049 0.06029273
#> [25,] 0.15792180 0.079710291 0.11844999 0.11225983 0.05773477 0.09837419
#>               V50          V51         V52         V53         V54
#>  [1,] 0.017373916  0.027005348 0.014598051 0.007549814 0.008302776
#>  [2,] 0.026588055  0.020811311 0.023405389 0.016576234 0.023748431
#>  [3,] 0.003417048  0.011969281 0.016305807 0.003739508 0.011599327
#>  [4,] 0.013517895  0.017347660 0.011787642 0.006336039 0.005741354
#>  [5,] 0.025631487  0.021645050 0.015744532 0.012323024 0.017053083
#>  [6,] 0.021707022  0.017294488 0.010760410 0.015441218 0.011240850
#>  [7,] 0.000535150 -0.003212603 0.004166966 0.010514264 0.008494246
#>  [8,] 0.018410245  0.021369159 0.016449757 0.010121132 0.011807373
#>  [9,] 0.012010014  0.012408785 0.012421438 0.011918798 0.009597542
#> [10,] 0.019615332  0.022987089 0.023285951 0.012419887 0.016345003
#> [11,] 0.027805014  0.039813472 0.027493764 0.010307184 0.010587958
#> [12,] 0.028482476  0.022576244 0.019588526 0.013559324 0.011072583
#> [13,] 0.019944865  0.020988807 0.017568221 0.019847681 0.021272929
#> [14,] 0.023364041  0.018145291 0.021629361 0.015212340 0.019240837
#> [15,] 0.016179618  0.009081989 0.004385280 0.011325760 0.007629646
#> [16,] 0.020796049  0.015510868 0.012477225 0.016505596 0.012735341
#> [17,] 0.077900000  0.039600000 0.017300000 0.014900000 0.011500000
#> [18,] 0.028700000  0.002700000 0.020800000 0.004800000 0.019900000
#> [19,] 0.009835855  0.005221558 0.018762438 0.026662138 0.016663784
#> [20,] 0.020256592  0.007157466 0.009573586 0.012882887 0.016461372
#> [21,] 0.029980686  0.017478255 0.011027891 0.007121395 0.012506765
#> [22,] 0.008290399  0.007110145 0.006644252 0.008709067 0.006051846
#> [23,] 0.007274985  0.004706605 0.008457970 0.008895946 0.008186890
#> [24,] 0.018188421  0.008658785 0.007571215 0.004665767 0.008397364
#> [25,] 0.009462182  0.017178984 0.010475118 0.009560488 0.009142069
#>                 V55         V56         V57         V58         V59         V6
#>  [1,]  8.827050e-03 0.005834366 0.004774624 0.006030853 0.006256955 0.14863666
#>  [2,]  2.142613e-02 0.013641652 0.011251709 0.014462492 0.012494793 0.11373237
#>  [3,] -9.321272e-05 0.004070637 0.010990501 0.006622014 0.003810324 0.15714775
#>  [4,]  4.484456e-03 0.003854928 0.003858816 0.002727823 0.003049682 0.08241721
#>  [5,]  1.592107e-02 0.006107575 0.007084972 0.014546387 0.010749331 0.16597047
#>  [6,]  5.366868e-03 0.012169804 0.006559277 0.012511041 0.007307050 0.14990749
#>  [7,]  3.839938e-03 0.014493227 0.009553706 0.007852934 0.012043302 0.15271895
#>  [8,]  6.157889e-03 0.006388127 0.003385647 0.006212146 0.007326747 0.13088982
#>  [9,]  1.681384e-03 0.002019163 0.001315658 0.004164861 0.007470824 0.09304652
#> [10,]  1.385744e-02 0.014617040 0.007537347 0.007529029 0.005838433 0.09037587
#> [11,]  1.010119e-02 0.009858079 0.006113795 0.012805189 0.014159126 0.08469195
#> [12,]  9.848698e-03 0.008923671 0.006053783 0.006374388 0.007020579 0.13463165
#> [13,]  1.428507e-02 0.004890653 0.002658874 0.016349550 0.005957951 0.02247883
#> [14,]  1.824031e-02 0.010826109 0.001876254 0.013312571 0.012918671 0.08567925
#> [15,]  9.509566e-03 0.007329631 0.007962966 0.007194139 0.004059191 0.15363397
#> [16,]  1.025535e-02 0.006998116 0.007738279 0.009011063 0.008650613 0.06400460
#> [17,]  2.020000e-02 0.013900000 0.002900000 0.016000000 0.010600000 0.10770000
#> [18,]  1.260000e-02 0.002200000 0.003700000 0.003400000 0.011400000 0.25870000
#> [19,]  2.057509e-02 0.005508994 0.004288093 0.005493230 0.010000501 0.06031970
#> [20,]  7.003872e-03 0.015696946 0.015207464 0.010368390 0.006273790 0.19400453
#> [21,]  6.357897e-03 0.006641596 0.009453446 0.007637115 0.007399652 0.08267674
#> [22,]  5.375582e-03 0.005691576 0.004597795 0.003096466 0.004826109 0.06997085
#> [23,]  4.234632e-03 0.002766293 0.002489422 0.003327106 0.001827869 0.02636985
#> [24,]  9.537940e-03 0.005441545 0.011569951 0.008395432 0.010452538 0.10682682
#> [25,]  9.097733e-03 0.009643717 0.012154902 0.013089263 0.013610960 0.11123866
#>               V60         V7         V8         V9
#>  [1,] 0.005402046 0.12196239 0.18673738 0.29564887
#>  [2,] 0.013417377 0.23959277 0.42517678 0.59082474
#>  [3,] 0.009661061 0.07695621 0.05950471 0.16660257
#>  [4,] 0.005909137 0.11701380 0.14880129 0.20861097
#>  [5,] 0.004148427 0.18512961 0.19384426 0.20968727
#>  [6,] 0.005275061 0.13381424 0.14483525 0.19605388
#>  [7,] 0.008265996 0.15652834 0.06875995 0.11532022
#>  [8,] 0.003708120 0.14941271 0.14562691 0.18472969
#>  [9,] 0.002098771 0.05716418 0.10744797 0.32477583
#> [10,] 0.009439954 0.11582838 0.05958374 0.15142123
#> [11,] 0.005503648 0.09788249 0.15375196 0.18112389
#> [12,] 0.006013431 0.14021355 0.15160683 0.27604774
#> [13,] 0.002088784 0.05234059 0.20512984 0.38880655
#> [14,] 0.007322038 0.11280113 0.10364790 0.10323062
#> [15,] 0.012595869 0.13484162 0.13766425 0.05247758
#> [16,] 0.001420891 0.07359257 0.07907254 0.06446606
#> [17,] 0.013400000 0.23630000 0.23850000 0.00750000
#> [18,] 0.007700000 0.23040000 0.20670000 0.34160000
#> [19,] 0.009349910 0.02968565 0.06998124 0.13871886
#> [20,] 0.004561807 0.13155311 0.12212197 0.10925466
#> [21,] 0.005075387 0.13039145 0.19269734 0.21862222
#> [22,] 0.004910838 0.07317069 0.07038321 0.10013693
#> [23,] 0.002778274 0.04524442 0.06322986 0.08494285
#> [24,] 0.008151471 0.12934653 0.12055549 0.09099422
#> [25,] 0.012126869 0.17059480 0.16950664 0.18374076
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478