Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

r format_bib("kohonen_1990", "kohonen_1995")

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                 V1         V10        V11        V12        V13        V14
#>  [1,]  0.009977462  0.14280098 0.21217207 0.30691831 0.39469878 0.47072728
#>  [2,]  0.085930205  0.37354730 0.33882209 0.21926738 0.23138999 0.33138903
#>  [3,]  0.030211917  0.01058586 0.05107900 0.06511262 0.11315089 0.26713210
#>  [4,]  0.034309584  0.36578456 0.36865967 0.33457319 0.29923701 0.17630801
#>  [5,]  0.039867959  0.17750443 0.19321241 0.27118818 0.32042087 0.28522174
#>  [6,]  0.025545404  0.31063542 0.37488355 0.41069004 0.41458679 0.46492036
#>  [7,]  0.032814959  0.36092478 0.44563458 0.45948714 0.41019684 0.33022940
#>  [8,]  0.028515878  0.28248714 0.32256631 0.36368912 0.30345559 0.21347372
#>  [9,]  0.048061586 -0.12613611 0.06935201 0.28349917 0.38166538 0.43626395
#> [10,]  0.089428532  0.43711291 0.50467919 0.51064110 0.64853731 0.82484886
#> [11,]  0.059897673  0.18546910 0.22917837 0.20466766 0.31041320 0.38107995
#> [12,]  0.017826699  0.23618980 0.32306887 0.31817282 0.27666724 0.22369433
#> [13,]  0.023423079  0.32665847 0.28392174 0.29291689 0.28241357 0.22574216
#> [14,]  0.018003360  0.27981239 0.23762833 0.11247284 0.20345207 0.49680354
#> [15,]  0.022562515  0.03805032 0.22152662 0.37735230 0.28115740 0.28386310
#> [16,]  0.022926532  0.15028720 0.15612298 0.21956026 0.30369077 0.14276286
#> [17,]  0.005920980  0.09640338 0.11030240 0.08119513 0.07689403 0.10482133
#> [18,]  0.023402627  0.11035274 0.14952640 0.07188890 0.13638026 0.17490793
#> [19,]  0.032039733  0.30838892 0.29952678 0.41823793 0.61130344 0.76493741
#> [20,]  0.020009806  0.02393716 0.07084286 0.20587898 0.17642342 0.20932693
#> [21,]  0.021922201  0.12625460 0.13543038 0.11164391 0.12332829 0.13017191
#> [22,]  0.029309619  0.07114153 0.04199171 0.07319695 0.11949220 0.12649953
#> [23,] -0.011137315  0.10009997 0.07971004 0.08489069 0.16262227 0.13934812
#> [24,]  0.019911036  0.02470155 0.14029742 0.38396092 0.32914715 0.29368482
#> [25,]  0.019579215  0.09982894 0.08987629 0.09263753 0.11918634 0.05108482
#>              V15        V16        V17       V18       V19           V2
#>  [1,] 0.44295419 0.42678054 0.54672391 0.6565367 0.7832242  0.025442227
#>  [2,] 0.42864251 0.41610263 0.29803171 0.2093381 0.3423871  0.085484554
#>  [3,] 0.40137859 0.60982130 0.71255546 0.6795330 0.8542792  0.033041600
#>  [4,] 0.16889224 0.20659788 0.19140616 0.2487887 0.2556396  0.048785533
#>  [5,] 0.38607353 0.50810998 0.55320457 0.5881755 0.5511088  0.063121504
#>  [6,] 0.48340928 0.51671193 0.62439121 0.6895210 0.7119264  0.005955612
#>  [7,] 0.26592704 0.27578882 0.32110082 0.3380102 0.4693555  0.033729616
#>  [8,] 0.18084120 0.21983053 0.22107219 0.3584710 0.5763402  0.031483233
#>  [9,] 0.55350986 0.57575933 0.74395825 0.7478424 0.7700953  0.056357359
#> [10,] 0.99677060 1.03603597 0.78090170 0.8342793 0.9297084  0.107742199
#> [11,] 0.35042184 0.42608779 0.46555623 0.3473655 0.4367284  0.088650783
#> [12,] 0.15180670 0.10708173 0.20002673 0.2533760 0.3632146  0.035528771
#> [13,] 0.18885647 0.22172745 0.21109419 0.2053960 0.2461898  0.025963168
#> [14,] 0.68546625 0.77218880 0.79731980 0.6547482 0.4273146  0.054487148
#> [15,] 0.29226996 0.41232229 0.60034699 0.4897789 0.4349567  0.057305025
#> [16,] 0.21824261 0.28063057 0.32122914 0.3673033 0.3542684  0.048382670
#> [17,] 0.16222081 0.14671699 0.14338545 0.1864486 0.2105359  0.004782023
#> [18,] 0.20894965 0.19567284 0.15138476 0.1428983 0.1122354  0.019646755
#> [19,] 0.88893822 0.88048502 0.81638938 0.7275313 0.5434686  0.042297531
#> [20,] 0.21945443 0.36689258 0.73165245 0.6284418 0.3087351  0.033906926
#> [21,] 0.15321460 0.25052616 0.31702735 0.3507337 0.3755744  0.018210469
#> [22,] 0.14802936 0.09311455 0.13828562 0.1830343 0.2595791  0.062796072
#> [23,] 0.14484760 0.29484971 0.34869652 0.3239839 0.3583814 -0.017614177
#> [24,] 0.04972616 0.08610221 0.22308570 0.4961031 0.3989495  0.022651990
#> [25,] 0.01022310 0.02728510 0.06948525 0.2054319 0.3662192  0.002774945
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.9072383 0.9957985 1.0176579 0.9414634 0.9000534 0.9123535 0.8601906
#>  [2,] 0.6269991 0.6620928 0.3510561 0.2433004 0.2089461 0.3135405 0.4247522
#>  [3,] 1.0208221 0.9953910 0.9808465 0.8301861 0.6749347 0.5523874 0.4706266
#>  [4,] 0.1791896 0.2737205 0.6248249 0.8353070 0.8582205 0.8452492 0.8357109
#>  [5,] 0.6367232 0.7181984 0.8366791 0.9743468 1.0168346 1.0192960 1.0273118
#>  [6,] 0.7897865 0.9261469 0.9602753 0.9936828 1.0661467 1.0455938 1.0291371
#>  [7,] 0.6068263 0.7228727 0.8105854 0.8164702 0.8454619 0.8795441 0.9165603
#>  [8,] 0.7107476 0.8132743 0.7957409 0.8379699 0.9482811 0.9160492 0.9648093
#>  [9,] 0.9483484 0.8750594 0.7905548 0.8679401 1.0506034 0.9202903 0.5632591
#> [10,] 1.0382784 0.9256074 0.8274759 0.9603176 0.8923771 0.6005803 0.4802103
#> [11,] 0.5712369 0.6118282 0.6815136 0.6529403 0.6464624 0.6509714 0.6965507
#> [12,] 0.3947738 0.4322419 0.4367531 0.4429965 0.5212050 0.6045238 0.7512885
#> [13,] 0.2558681 0.2464888 0.2716194 0.3975788 0.4966677 0.6127979 0.7318564
#> [14,] 0.3410426 0.5516886 0.7595541 0.8564620 0.9286591 0.8092212 0.7123559
#> [15,] 0.6222329 0.3557679 0.1830144 0.3709366 0.9861458 1.0042657 0.4783466
#> [16,] 0.1537501 0.2160955 0.1976543 0.2112044 0.2477386 0.2963134 0.3092385
#> [17,] 0.2515039 0.3198762 0.3985504 0.5371527 0.7351569 0.9017276 0.8839946
#> [18,] 0.1775891 0.2881211 0.4985288 0.7598925 0.8305892 0.7998090 0.8257163
#> [19,] 0.4892175 0.2651009 0.4072079 0.5805566 0.4332923 0.3019630 0.5637317
#> [20,] 0.2192671 0.3826130 0.6033069 0.6888717 0.7382688 0.8508248 0.9103831
#> [21,] 0.4167626 0.5005430 0.5203895 0.5350884 0.5822035 0.5906952 0.6701999
#> [22,] 0.2455584 0.1743178 0.2248424 0.2857923 0.3704685 0.4500059 0.3630248
#> [23,] 0.3923818 0.4309020 0.4304041 0.4737373 0.4655918 0.6141389 0.8825897
#> [24,] 0.6490981 0.7321114 0.7551811 0.8183234 0.8626673 1.0084150 0.8590710
#> [25,] 0.3976484 0.3660261 0.4392250 0.5890970 0.7187888 0.8006802 0.7958088
#>             V27       V28       V29           V3       V30       V31        V32
#>  [1,] 0.7421111 0.6532124 0.5493010  0.026675947 0.3822214 0.1982304 0.08336587
#>  [2,] 0.3572957 0.6167065 0.6332264  0.065150896 0.6902156 0.5387070 0.52545574
#>  [3,] 0.4381230 0.3125809 0.4348930  0.035818094 0.4537873 0.2203759 0.17674663
#>  [4,] 0.8640121 0.9390384 0.7955067  0.065512848 0.5500232 0.6352746 0.61156660
#>  [5,] 0.9837340 1.0407321 0.8566681  0.001015819 0.5948491 0.5619112 0.55383748
#>  [6,] 1.0373375 0.9828498 0.7900450  0.024676118 0.6150486 0.4640110 0.38843223
#>  [7,] 0.9105122 0.8036519 0.6499975  0.038876289 0.5094723 0.3800622 0.38992621
#>  [8,] 0.9880107 0.9550682 0.8483751  0.042314052 0.6557040 0.4381462 0.27757004
#>  [9,] 0.3088757 0.2426022 0.2872097  0.093834819 0.2440560 0.3020944 0.34914151
#> [10,] 0.5368378 0.5245632 0.4416570  0.065911952 0.2953428 0.2621068 0.17200970
#> [11,] 0.7430686 0.7755106 0.7246058  0.109116013 0.8095112 0.8467983 0.82141842
#> [12,] 0.9634115 1.0223933 0.9345013  0.033621252 0.7529127 0.6087681 0.55021669
#> [13,] 0.7770829 0.8476704 0.8738015  0.034080930 0.9206327 0.9043182 0.76099493
#> [14,] 0.7119724 0.5713854 0.3917667  0.092763509 0.4348088 0.4778634 0.28989999
#> [15,] 0.2822630 0.6657162 0.8232770  0.068623352 0.3948762 0.5318034 0.72071525
#> [16,] 0.1910273 0.4891928 0.7069134  0.041503911 0.6899225 0.6052223 0.61749273
#> [17,] 0.7968835 0.7216707 0.5785083 -0.006391479 0.4074072 0.2659687 0.22539603
#> [18,] 0.8089483 0.8700946 0.9460167  0.029244084 0.8129736 0.6704212 0.55130229
#> [19,] 0.6262165 0.5906159 0.5627784  0.016341374 0.5615672 0.6281767 0.52643058
#> [20,] 0.8791820 0.7883715 0.8575697  0.041241561 0.8906116 0.8402338 0.73265258
#> [21,] 0.7359782 0.7739935 0.8123299  0.024741914 0.8066993 0.7260142 0.64437568
#> [22,] 0.2702868 0.4072350 0.5954381  0.033722642 0.7882348 0.8161696 0.75927817
#> [23,] 0.9089746 0.6545647 0.4056830 -0.002841966 0.3983259 0.4821681 0.47264268
#> [24,] 0.6152790 0.4634785 0.4308904  0.068337669 0.4210026 0.5068959 0.52141970
#> [25,] 0.7000270 0.7894223 0.9691556  0.012172458 0.9730421 0.9337033 0.89658607
#>               V33         V34         V35        V36        V37        V38
#>  [1,]  0.06330300  0.03919624  0.14104445 0.32675581 0.28028386 0.31887676
#>  [2,]  0.55918650  0.66946205  0.72922518 0.62416856 0.49702260 0.62316631
#>  [3,]  0.38444333  0.36669986  0.35617215 0.34182831 0.38809787 0.37750378
#>  [4,]  0.49444533  0.22823663  0.09138718 0.07485914 0.20779047 0.19561011
#>  [5,]  0.36902498  0.14038131  0.03996674 0.08284257 0.25577590 0.41035172
#>  [6,]  0.16744166 -0.05381523 -0.07445213 0.13749975 0.27033864 0.23992461
#>  [7,]  0.37410612  0.31538840  0.17768162 0.24239479 0.27524563 0.31181403
#>  [8,]  0.19766133  0.17517042  0.13129948 0.02879440 0.06121212 0.08500623
#>  [9,]  0.41069713  0.48378828  0.53172227 0.33756429 0.25506213 0.40138684
#> [10,] -0.01833264  0.04742885  0.12305000 0.17290151 0.15983660 0.11506835
#> [11,]  0.58481049  0.48481212  0.46871326 0.22637423 0.23773908 0.35710620
#> [12,]  0.48996381  0.41628687  0.23822407 0.13104519 0.23728188 0.16793330
#> [13,]  0.61970630  0.51888059  0.45674353 0.47733548 0.50344614 0.46848124
#> [14,]  0.20963394  0.13551370  0.45381344 0.68797534 0.68358795 0.38992296
#> [15,]  0.56076977  0.89061487  0.76659745 0.69413219 0.74517187 0.39490491
#> [16,]  0.64863111  0.61386966  0.61885489 0.75996685 0.83207195 0.90405024
#> [17,]  0.21322016  0.24380162  0.23937118 0.20716308 0.20561797 0.19512213
#> [18,]  0.57676822  0.66151197  0.62510902 0.48969261 0.25782138 0.14603272
#> [19,]  0.29258216  0.21327843  0.18774107 0.38513606 0.40300426 0.36103934
#> [20,]  0.70800434  0.80402797  0.95382583 1.04100055 0.94332175 0.76941022
#> [21,]  0.65593683  0.65135187  0.61745692 0.59884491 0.54229837 0.45534141
#> [22,]  0.79814165  0.88375740  0.96758461 0.92088519 0.64009198 0.41515814
#> [23,]  0.39013834  0.43106227  0.46963092 0.47104643 0.44245949 0.38223325
#> [24,]  0.60195941  0.57529344  0.57727382 0.55153557 0.42885205 0.19959729
#> [25,]  0.93928826  0.91006697  0.73919455 0.50453092 0.29846143 0.17900815
#>              V39           V4       V40        V41        V42        V43
#>  [1,] 0.44556458  0.061431543 0.3715430 0.32572542 0.34984310 0.24327118
#>  [2,] 0.54306044  0.097899406 0.3465287 0.44003136 0.53992090 0.48861502
#>  [3,] 0.32020606  0.048076882 0.3882040 0.35990693 0.38870250 0.39937103
#>  [4,] 0.17746557  0.093894661 0.1226171 0.18591847 0.10961234 0.10209522
#>  [5,] 0.36604465  0.016563769 0.3743320 0.33058824 0.28567308 0.21646173
#>  [6,] 0.23983100  0.098787258 0.3304177 0.27006280 0.33353626 0.34235149
#>  [7,] 0.40610110  0.039455151 0.3721526 0.21232159 0.23848359 0.30667340
#>  [8,] 0.09429806  0.047513453 0.1153944 0.05692523 0.11415193 0.21842502
#>  [9,] 0.42603834  0.043731603 0.3798834 0.25028298 0.25845602 0.24251485
#> [10,] 0.22260846  0.009132220 0.2822581 0.21212300 0.09894642 0.14430235
#> [11,] 0.36313342  0.118818496 0.1944343 0.52729120 0.64274920 0.43702347
#> [12,] 0.13410828  0.056933543 0.3064070 0.38129450 0.39594224 0.30989572
#> [13,] 0.47980108  0.044555679 0.5383510 0.51272190 0.44372744 0.45942720
#> [14,] 0.19371458  0.076663868 0.3394882 0.38822941 0.33824952 0.24422315
#> [15,] 0.26120908  0.092874457 0.3012903 0.28755664 0.29793180 0.04206185
#> [16,] 0.94339069  0.026539877 0.8902098 0.84608829 0.74831889 0.58300178
#> [17,] 0.21279577 -0.004870306 0.2196024 0.23896568 0.19679285 0.16014127
#> [18,] 0.17052048  0.043161556 0.2174674 0.18421885 0.20849593 0.29056345
#> [19,] 0.36788765  0.004429514 0.3552463 0.26811722 0.22329160 0.10506735
#> [20,] 0.60529787  0.023808688 0.3896438 0.26997105 0.19814905 0.23505116
#> [21,] 0.30352211  0.026293522 0.2573031 0.23187167 0.18419746 0.15148896
#> [22,] 0.32922950  0.015661391 0.3047139 0.34349071 0.37071044 0.21057745
#> [23,] 0.47188542 -0.029121922 0.4706660 0.31992627 0.22524133 0.18358264
#> [24,] 0.05810670  0.052355595 0.3907392 0.42095178 0.54052934 0.43237763
#> [25,] 0.10934797  0.010573726 0.2640828 0.32458095 0.35090630 0.31144676
#>              V44         V45         V46         V47         V48          V49
#>  [1,] 0.21003243  0.15778319  0.14320669  0.14697532  0.11214897  0.077301384
#>  [2,] 0.41428722  0.50265873  0.49449486  0.29787346  0.21373633  0.106910719
#>  [3,] 0.29093989  0.13638129  0.09892312  0.13380789  0.08411435  0.039671692
#>  [4,] 0.13263871  0.06917780  0.07876551  0.09875473  0.13094084  0.084711670
#>  [5,] 0.24699663  0.18177258  0.12912584  0.11707458  0.08339110  0.046249613
#>  [6,] 0.20035016  0.05339323  0.10443032  0.19016622  0.14997701  0.060026588
#>  [7,] 0.33679453  0.32015529  0.20628796  0.11919251  0.07781745  0.051236561
#>  [8,] 0.23888302  0.22051663  0.23563784  0.18833650  0.10678639  0.058120824
#>  [9,] 0.23594827  0.24989238  0.19641976  0.08955605  0.03509152  0.036306766
#> [10,] 0.27080842  0.12219271  0.02373433  0.06771825  0.07843422  0.069811747
#> [11,] 0.40374191  0.58424717  0.54708405  0.29337531  0.16183164  0.092827731
#> [12,] 0.14747388  0.19534272  0.15422350  0.17498053  0.16088186  0.084234425
#> [13,] 0.49282891  0.54119836  0.41592745  0.29154755  0.22337020  0.124425256
#> [14,] 0.20219248  0.20499833  0.15001917  0.15503848  0.14056146  0.100588366
#> [15,] 0.17743418  0.12929115  0.14980232  0.12219437  0.07659785  0.046287928
#> [16,] 0.38568541  0.40556320  0.44779166  0.35713118  0.23979913  0.154172702
#> [17,] 0.15965449  0.13801374  0.10806619  0.03437599  0.03307846  0.025950021
#> [18,] 0.18025807  0.05395639  0.05362937  0.05252796  0.04547754  0.011609800
#> [19,] 0.11359768  0.10288816  0.09650759  0.05043951  0.03366483  0.027354003
#> [20,] 0.15069214  0.18756919  0.18193372  0.10800629  0.01762867  0.004573619
#> [21,] 0.15057612  0.11053682  0.09491685  0.05208239  0.04118408  0.022928366
#> [22,] 0.24355042  0.18086145  0.14478145  0.14851501  0.10903747  0.055503194
#> [23,] 0.09638716 -0.03083857 -0.01552589 -0.01499810 -0.02494965 -0.006121669
#> [24,] 0.22994742  0.11911079  0.07924551  0.04714275  0.06681315  0.027364785
#> [25,] 0.31036477  0.29154551  0.20578357  0.09823610  0.03244196  0.025652481
#>                 V5          V50           V51         V52           V53
#>  [1,]  0.165900579 0.0307860904  0.0219564751 0.020562819  0.0158582935
#>  [2,]  0.084729401 0.0372999537  0.0529893288 0.043771992  0.0257906696
#>  [3,]  0.077788981 0.0191879097  0.0123122711 0.006016386  0.0194835198
#>  [4,]  0.119474301 0.0267226441  0.0176939138 0.014206745  0.0072133443
#>  [5,]  0.040357340 0.0012791597  0.0161841382 0.000796841  0.0129093469
#>  [6,]  0.152040491 0.0220189447  0.0140880348 0.002453490  0.0088884390
#>  [7,]  0.064806859 0.0153861670  0.0185781963 0.009989880  0.0075833872
#>  [8,]  0.075013931 0.0162270814  0.0196148910 0.014746066  0.0048522545
#>  [9,]  0.107169561 0.0009405305 -0.0115871187 0.007073994  0.0195074843
#> [10,]  0.054368448 0.0058591434  0.0327965952 0.017577669  0.0151230162
#> [11,]  0.147964652 0.0449488452  0.0319117009 0.024509825  0.0227303129
#> [12,]  0.075833822 0.0110097398  0.0217619858 0.015027892  0.0096280563
#> [13,]  0.070940964 0.0360052722  0.0149425737 0.014617813  0.0124878015
#> [14,]  0.053593436 0.0302204322  0.0155932164 0.011821542  0.0063220940
#> [15,]  0.169493502 0.0050631398  0.0152340449 0.001726261  0.0156964482
#> [16,]  0.071281786 0.0578457103  0.0310706186 0.014709020  0.0154621372
#> [17,]  0.011011584 0.0071217731  0.0005630555 0.005088641  0.0044481000
#> [18,]  0.050500920 0.0100887193  0.0112183301 0.007440752  0.0078435413
#> [19,]  0.051105160 0.0085345488  0.0086785190 0.005689475 -0.0004318899
#> [20,]  0.002905937 0.0355714984  0.0081992688 0.009590174  0.0044270522
#> [21,]  0.055788135 0.0149785124  0.0072297701 0.008409682  0.0063529685
#> [22,]  0.043097764 0.0117242533  0.0142072361 0.005628995  0.0104179205
#> [23,] -0.031182870 0.0031523186  0.0067756617 0.004025762  0.0056021537
#> [24,]  0.059969882 0.0079647324  0.0014795815 0.006194696  0.0134979227
#> [25,]  0.015355891 0.0128557172  0.0081778727 0.008950500  0.0069789151
#>               V54          V55          V56         V57         V58         V59
#>  [1,] 0.024882483  0.012762294 0.0062933019 0.009665824 0.010339101 0.012755392
#>  [2,] 0.026806837  0.019970297 0.0150304410 0.015689549 0.017530445 0.017816836
#>  [3,] 0.017288313 -0.004101865 0.0140269106 0.004346612 0.002208115 0.015073143
#>  [4,] 0.007369687  0.005972230 0.0107155150 0.007292478 0.007258289 0.008203847
#>  [5,] 0.011115816  0.012283681 0.0108298271 0.004949784 0.002013509 0.008474456
#>  [6,] 0.017461843 -0.001014782 0.0173452053 0.004561680 0.032859094 0.009107662
#>  [7,] 0.008243221  0.008017329 0.0064728811 0.004980155 0.006314423 0.006906228
#>  [8,] 0.002538481  0.002945702 0.0037330783 0.003764201 0.002195524 0.001441165
#>  [9,] 0.007144513  0.006503341 0.0182538207 0.010995825 0.009001117 0.013732225
#> [10,] 0.013514040  0.024577477 0.0052012241 0.007529560 0.016620660 0.011011697
#> [11,] 0.018534035  0.022404913 0.0184383987 0.016827743 0.018233740 0.018155600
#> [12,] 0.009808464  0.003355912 0.0058431430 0.002054754 0.007053467 0.005872932
#> [13,] 0.007552146  0.005303761 0.0064100597 0.006611493 0.006041953 0.004834596
#> [14,] 0.013587626  0.006903997 0.0061646032 0.008420652 0.006771095 0.007632494
#> [15,] 0.005597185  0.020609938 0.0080056679 0.015826941 0.004226960 0.021087928
#> [16,] 0.009160501  0.015530343 0.0121003234 0.002305614 0.016768347 0.013124595
#> [17,] 0.006329222  0.002496750 0.0009021234 0.001684179 0.003719874 0.001307876
#> [18,] 0.005981437  0.005060420 0.0063057537 0.011437980 0.009578026 0.007765555
#> [19,] 0.004596072  0.002343754 0.0075769986 0.003610545 0.002912762 0.005458350
#> [20,] 0.013928908  0.011981377 0.0034311231 0.010250788 0.006382893 0.007856715
#> [21,] 0.010661193  0.010520982 0.0087704503 0.008244001 0.005440989 0.005247548
#> [22,] 0.007700013  0.009797553 0.0081119452 0.008608733 0.013795069 0.009138233
#> [23,] 0.009167571  0.010537845 0.0064852357 0.006425432 0.003232535 0.004260944
#> [24,] 0.012927234  0.011808157 0.0156123540 0.009702214 0.004334243 0.005353667
#> [25,] 0.007985664  0.002460439 0.0009237983 0.001792758 0.002585517 0.006004127
#>               V6          V60          V7           V8          V9
#>  [1,] 0.15120472  0.007212398 0.154715860 0.1528533134  0.14731714
#>  [2,] 0.11429050  0.012234708 0.137342341 0.2858712557  0.38866333
#>  [3,] 0.07867557  0.007682023 0.003801834 0.0967200685  0.05260441
#>  [4,] 0.13369772  0.009467457 0.152499255 0.2299851705  0.28185911
#>  [5,] 0.07215914  0.003058049 0.092155721 0.1288646421  0.13738452
#>  [6,] 0.17921819  0.008281397 0.128549033 0.1460729245  0.24882788
#>  [7,] 0.09852692  0.006016661 0.123632708 0.1691776153  0.28095849
#>  [8,] 0.09363854  0.004979213 0.128217647 0.1613839449  0.21171803
#>  [9,] 0.14384371  0.003819586 0.174441436 0.0007001247 -0.01153360
#> [10,] 0.19331812  0.015404269 0.219116161 0.2194857187  0.31521932
#> [11,] 0.11091724  0.014604966 0.100160859 0.1014089169  0.20481048
#> [12,] 0.08934787  0.003311959 0.095666755 0.0863390719  0.18635330
#> [13,] 0.10981674  0.003395813 0.124111325 0.1426563337  0.27178859
#> [14,] 0.10455471  0.004728909 0.143876409 0.1997517414  0.23532574
#> [15,] 0.24017786  0.006654619 0.207034194 0.1797251617  0.10840338
#> [16,] 0.10545050  0.012964409 0.190834585 0.2020572145  0.04839769
#> [17,] 0.02227129  0.002878041 0.062666127 0.0861537472  0.08899095
#> [18,] 0.08222423  0.006395726 0.068834736 0.1024702731  0.09754676
#> [19,] 0.15544654  0.004931130 0.179828654 0.2667094673  0.27949081
#> [20,] 0.04741363  0.009481961 0.130156199 0.1111566105  0.02117483
#> [21,] 0.10542448  0.004231139 0.112780216 0.0962052686  0.13614672
#> [22,] 0.07550904  0.002535980 0.087530154 0.0315509135  0.04953970
#> [23,] 0.05413034 -0.003929614 0.036318557 0.0691560648  0.06755301
#> [24,] 0.11487841  0.014457940 0.109253149 0.1588884167  0.05280865
#> [25,] 0.01029672  0.003576278 0.057753642 0.0297435082  0.04370397
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768