Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

r format_bib("kohonen_1990", "kohonen_1995")

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1          V10        V11        V12       V13       V14
#>  [1,] 0.004667077  0.132468174 0.22207971 0.25951910 0.2346800 0.1474618
#>  [2,] 0.017230163  0.002076441 0.22269598 0.26343902 0.1654283 0.3740599
#>  [3,] 0.013100000  0.319300000 0.34680000 0.37380000 0.3055000 0.1926000
#>  [4,] 0.027197895  0.346590127 0.42014192 0.43864710 0.4570613 0.3801449
#>  [5,] 0.083580608  0.123399881 0.23655391 0.29449035 0.4455111 0.5938198
#>  [6,] 0.061647960  0.390074192 0.34798147 0.26158124 0.2249128 0.2505852
#>  [7,] 0.024519267  0.204791983 0.25210849 0.29503696 0.2845830 0.2672501
#>  [8,] 0.133737523  0.213243737 0.28519235 0.21841048 0.3086453 0.3265150
#>  [9,] 0.009741949  0.227409675 0.29880102 0.30514767 0.2840553 0.2393428
#> [10,] 0.022752876  0.196551436 0.23207506 0.27100586 0.3138414 0.3719175
#> [11,] 0.028198269  0.192725785 0.23052948 0.27483129 0.2533322 0.2359357
#> [12,] 0.046933104  0.289723518 0.35432652 0.49110903 0.5363998 0.6277942
#> [13,] 0.032300000  0.215400000 0.30850000 0.34250000 0.2990000 0.1402000
#> [14,] 0.081935939  0.156686241 0.23912362 0.26641281 0.3084096 0.3095401
#> [15,] 0.028899827  0.096903239 0.04376597 0.04224129 0.1512741 0.3218609
#> [16,] 0.016989779  0.104694202 0.15915955 0.12778872 0.1936691 0.2660538
#> [17,] 0.022562903  0.045072005 0.08116839 0.15631663 0.2391890 0.4118553
#> [18,] 0.017475290  0.151323563 0.17528524 0.21034535 0.2900359 0.2779699
#> [19,] 0.017531242  0.100180637 0.10558407 0.12404985 0.1396463 0.1557061
#> [20,] 0.022187074  0.131800045 0.14457763 0.23720797 0.1677626 0.1862801
#> [21,] 0.017317337  0.005700041 0.03338747 0.14214273 0.1763558 0.2088323
#> [22,] 0.024726534  0.283104142 0.25578481 0.13988053 0.2304546 0.4802555
#> [23,] 0.009953309  0.080472857 0.13350395 0.13383978 0.2067340 0.2643754
#> [24,] 0.008639032  0.057627815 0.03785264 0.07314877 0.2285532 0.2312327
#> [25,] 0.011107913 -0.010848864 0.05043404 0.14903949 0.2356067 0.3558913
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.03078148 0.1622948 0.3308703 0.4416325 0.5249659 0.014450691 0.6116967
#>  [2,] 0.46248443 0.5695990 0.4678722 0.3792648 0.3509920 0.040532009 0.2480485
#>  [3,] 0.13850000 0.2122000 0.2758000 0.4576000 0.6487000 0.038700000 0.7154000
#>  [4,] 0.24953288 0.3226973 0.3685241 0.3732121 0.5421309 0.035001825 0.7134990
#>  [5,] 0.64695844 0.6700101 0.7152633 0.7737224 0.8623569 0.085659830 0.8749708
#>  [6,] 0.32745943 0.3592478 0.2127468 0.1761861 0.2381167 0.071037433 0.4697612
#>  [7,] 0.21327524 0.1738445 0.1829998 0.2533444 0.3315086 0.030519360 0.3928437
#>  [8,] 0.36216013 0.3895394 0.4256825 0.4602625 0.5397580 0.241301045 0.3695141
#>  [9,] 0.18588771 0.1865322 0.2298000 0.3667141 0.6221592 0.053933290 0.7261408
#> [10,] 0.42990028 0.5310074 0.6167293 0.6965838 0.7627472 0.030569722 0.8299988
#> [11,] 0.18222735 0.2108746 0.2190845 0.3103776 0.4949411 0.035228842 0.6346081
#> [12,] 0.69397317 0.7791409 0.9379392 0.9017734 0.9622785 0.062391914 0.9029777
#> [13,] 0.12350000 0.1534000 0.1901000 0.2429000 0.2120000 0.010100000 0.2395000
#> [14,] 0.34981229 0.3424937 0.3937795 0.1974060 0.3350096 0.089171432 0.3585146
#> [15,] 0.51801856 0.7201096 0.7944772 0.7960954 0.8253240 0.021890385 0.8453706
#> [16,] 0.25425595 0.2080128 0.2618022 0.3384444 0.4396801 0.009592801 0.4731447
#> [17,] 0.53331526 0.6991543 0.7731782 0.7946896 0.7824959 0.033940729 0.8603278
#> [18,] 0.33322560 0.4305748 0.4545822 0.3628151 0.2057579 0.021853781 0.2399882
#> [19,] 0.16562394 0.2073260 0.2061683 0.1902515 0.2710020 0.027785300 0.3029244
#> [20,] 0.19719601 0.2658938 0.2385464 0.3565801 0.4451029 0.034603427 0.3383511
#> [21,] 0.16590978 0.1530355 0.2549004 0.3388607 0.3141192 0.023638111 0.5407083
#> [22,] 0.62429693 0.7217088 0.7730635 0.6826116 0.5059618 0.063836089 0.3999557
#> [23,] 0.27155530 0.1965976 0.2379948 0.3112332 0.4063747 0.002473080 0.4694952
#> [24,] 0.17570372 0.2461034 0.2436205 0.3924547 0.7049559 0.028646465 0.8424482
#> [25,] 0.38719968 0.3695253 0.4252551 0.6586777 0.7388572 0.013513569 0.7765509
#>             V21        V22       V23       V24       V25        V26       V27
#>  [1,] 0.7233947 0.83172213 0.8632247 0.8274561 0.8852182 1.00396842 1.0221875
#>  [2,] 0.2204765 0.09296025 0.2516146 0.5266734 0.3856831 0.41441186 0.5294388
#>  [3,] 0.8010000 0.79240000 0.8793000 1.0000000 0.9865000 0.94740000 0.9474000
#>  [4,] 0.7652039 0.82238770 0.8125003 0.8008683 0.8429954 0.90966470 0.9013830
#>  [5,] 0.8876464 0.89005034 0.9547545 0.9147260 0.6642882 0.58057896 0.4681293
#>  [6,] 0.5012804 0.34336503 0.2979971 0.3857117 0.4742704 0.60022060 0.6150040
#>  [7,] 0.4404133 0.44040776 0.5326162 0.6528214 0.7356247 0.78408352 0.8784949
#>  [8,] 0.2158726 0.18704117 0.4275376 0.6582122 0.5154451 0.28186057 0.4037718
#>  [9,] 0.8049987 0.81077796 0.8151175 0.9239158 0.9378519 0.96292840 0.9992680
#> [10,] 0.8800175 0.91834293 0.9478853 0.9404376 0.8922008 0.88088707 0.8928371
#> [11,] 0.7679571 0.80521243 0.7866994 0.8807907 0.9122099 0.94825370 0.9665633
#> [12,] 0.9140273 0.84762038 0.6266890 0.3547660 0.2107987 0.09082559 0.1845922
#> [13,] 0.3272000 0.59490000 0.8302000 0.9045000 0.9888000 0.99120000 0.9448000
#> [14,] 0.4044683 0.39497035 0.4302391 0.3935088 0.4476450 0.47136951 0.5412482
#> [15,] 0.9967879 0.98029595 0.9683067 0.9959993 0.9569133 0.97016380 0.9026892
#> [16,] 0.5306371 0.59063473 0.6341559 0.7235819 0.9011897 0.94398891 1.0053999
#> [17,] 0.9875670 0.88253044 0.7883976 0.8511675 0.8962118 0.93697071 0.8652668
#> [18,] 0.3840586 0.41587357 0.4377726 0.6492655 0.8233917 0.76942779 0.7394000
#> [19,] 0.3069845 0.34968850 0.4212236 0.4448257 0.4581406 0.53966725 0.5852937
#> [20,] 0.2712920 0.29412892 0.3894522 0.3575993 0.2991995 0.42679208 0.3760142
#> [21,] 0.6126885 0.66026752 0.7154956 0.7666871 0.8981134 0.89954288 0.9520220
#> [22,] 0.5496826 0.73221393 0.8106666 0.8596927 0.7490873 0.64986505 0.6493839
#> [23,] 0.4989975 0.55620507 0.5951933 0.6461765 0.8223628 0.89249968 0.9498357
#> [24,] 0.7568320 0.63597224 0.4312207 0.4276169 0.6290220 0.75435752 0.8962419
#> [25,] 0.9057108 0.96718582 0.9886192 0.9806777 0.8882696 0.76355702 0.6240929
#>             V28        V29          V3       V30        V31       V32
#>  [1,] 0.9702554 0.86795094 0.028322274 0.6316780 0.22343539 0.2697485
#>  [2,] 0.4731756 0.41334010 0.054754811 0.3973930 0.36851692 0.3921684
#>  [3,] 0.9315000 0.83260000 0.032900000 0.6213000 0.37720000 0.2822000
#>  [4,] 0.8117258 0.63222944 0.043111940 0.5165894 0.47905329 0.5649783
#>  [5,] 0.4145692 0.38300432 0.090096551 0.2578547 0.19075122 0.2145320
#>  [6,] 0.7317154 0.70082244 0.050456861 0.7987510 0.65007504 0.6247380
#>  [7,] 0.9517613 0.92613113 0.034058962 0.7275486 0.61120871 0.4758645
#>  [8,] 0.5770835 0.51526389 0.317304599 0.4153846 0.36525532 0.2238569
#>  [9,] 0.9869237 0.87830960 0.033969228 0.6388883 0.39575868 0.3147971
#> [10,] 0.8494790 0.71761407 0.033209782 0.5591916 0.38366403 0.2900410
#> [11,] 0.9234110 0.80549135 0.042920922 0.6103833 0.41912014 0.3175294
#> [12,] 0.1710145 0.07041116 0.041798924 0.1112287 0.01562823 0.1037126
#> [13,] 1.0000000 0.90920000 0.029800000 0.7412000 0.76910000 0.7117000
#> [14,] 0.7719689 0.80162976 0.058926769 0.7422028 0.79044304 0.6944234
#> [15,] 0.6714726 0.37244718 0.044967519 0.3828646 0.46657676 0.3841217
#> [16,] 0.9887019 0.85266457 0.009617605 0.7465044 0.65371471 0.5037647
#> [17,] 0.6292749 0.31407527 0.001901654 0.2732922 0.32223265 0.2437777
#> [18,] 0.8420248 0.75470282 0.013210820 0.5346117 0.45864514 0.3657946
#> [19,] 0.7114098 0.76364905 0.031126005 0.7486232 0.68564558 0.5956395
#> [20,] 0.3580673 0.47389316 0.058791814 0.5679701 0.68338860 0.7097302
#> [21,] 0.8784154 0.69294149 0.026268187 0.6103084 0.55780770 0.4612010
#> [22,] 0.5352903 0.37944897 0.096654033 0.4348084 0.44337081 0.2839688
#> [23,] 0.9328042 0.77275926 0.010654863 0.6930872 0.65883743 0.4968072
#> [24,] 0.8079792 0.56288327 0.021183319 0.4212342 0.40875859 0.3666764
#> [25,] 0.5039779 0.56469113 0.004803413 0.5766130 0.48810642 0.2205727
#>              V33       V34       V35        V36         V37        V38
#>  [1,] 0.34757081 0.4892211 0.7652345 0.92980869  0.83512787 0.63353109
#>  [2,] 0.47739174 0.4013080 0.2067647 0.15169518  0.06086376 0.18459465
#>  [3,] 0.20420000 0.2190000 0.2223000 0.13270000  0.05210000 0.06180000
#>  [4,] 0.51474391 0.3936894 0.1709996 0.26679916  0.22601899 0.33529622
#>  [5,] 0.25257061 0.2767255 0.3409137 0.33066163  0.31421975 0.35345068
#>  [6,] 0.61305004 0.6606122 0.7685429 0.75570744  0.69965304 0.68041940
#>  [7,] 0.42836574 0.3282069 0.1804285 0.13063081  0.21549495 0.18632284
#>  [8,] 0.07737274 0.1457173 0.2937805 0.45945183  0.36861147 0.33788569
#>  [9,] 0.26111893 0.2199043 0.2165599 0.12262173  0.07062316 0.07685655
#> [10,] 0.19830402 0.1348002 0.1872816 0.20950725  0.24382909 0.27976613
#> [11,] 0.27419086 0.2241074 0.1540544 0.05822024  0.09273267 0.15001122
#> [12,] 0.25203121 0.2146558 0.3189427 0.07962830  0.23144319 0.32246862
#> [13,] 0.53040000 0.2131000 0.0928000 0.12970000  0.11590000 0.12260000
#> [14,] 0.47870909 0.5569567 0.6254160 0.38807833  0.16881042 0.31813212
#> [15,] 0.30737876 0.2754604 0.1805664 0.05615682  0.17674944 0.14943729
#> [16,] 0.37087036 0.2076735 0.1235643 0.27792680  0.26334218 0.31975832
#> [17,] 0.24060533 0.1664496 0.1018521 0.03373633  0.10788111 0.13374552
#> [18,] 0.22267799 0.1983300 0.3008764 0.29251547  0.17129248 0.15018647
#> [19,] 0.64679490 0.7270605 0.7677844 0.72636135  0.55940687 0.40401903
#> [20,] 0.62031701 0.4919383 0.5322790 0.53694368  0.68976868 0.91535057
#> [21,] 0.38193326 0.2624304 0.2914803 0.35961776  0.32913616 0.30493009
#> [22,] 0.22657489 0.1384021 0.3970040 0.62830200  0.65884195 0.39953944
#> [23,] 0.34812822 0.2130121 0.1246006 0.20153174  0.23404585 0.27823963
#> [24,] 0.36686531 0.2723007 0.3053413 0.40283196  0.53155896 0.48422342
#> [25,] 0.19301006 0.2386948 0.1528938 0.12764127 -0.02022526 0.04695368
#>              V39          V4       V40        V41         V42        V43
#>  [1,] 0.44557459 0.033445590 0.3084545 0.31311449 0.263963643 0.35355668
#>  [2,] 0.31710525 0.081006615 0.3767727 0.29801115 0.319003385 0.27365085
#>  [3,] 0.14160000 0.007800000 0.1460000 0.08460000 0.105500000 0.16390000
#>  [4,] 0.54243525 0.053512990 0.4493889 0.25892220 0.354798988 0.40595019
#>  [5,] 0.37485243 0.101518196 0.2922064 0.20456808 0.285974823 0.32337459
#>  [6,] 0.60455407 0.075569389 0.4912555 0.43225123 0.405630498 0.38081111
#>  [7,] 0.23827438 0.042292160 0.2586300 0.27377016 0.338903318 0.23253234
#>  [8,] 0.38461322 0.443520170 0.4258461 0.36748428 0.272281159 0.16444178
#>  [9,] 0.12087143 0.028834264 0.1179683 0.04757241 0.063763218 0.15978994
#> [10,] 0.27383371 0.053101894 0.2716562 0.27331163 0.269471316 0.22229420
#> [11,] 0.12134591 0.034089688 0.1329951 0.06999995 0.141103843 0.22125670
#> [12,] 0.20599597 0.125330596 0.1302813 0.20044306 0.339198391 0.24929253
#> [13,] 0.17680000 0.056400000 0.0345000 0.15620000 0.082400000 0.11490000
#> [14,] 0.28421389 0.055759740 0.1690025 0.55254849 0.762328575 0.67029354
#> [15,] 0.09033202 0.045919667 0.1534926 0.08893477 0.043375793 0.04406991
#> [16,] 0.35767162 0.024716038 0.2240883 0.10952669 0.262102422 0.20538890
#> [17,] 0.12703338 0.025456794 0.1504851 0.11594138 0.047782377 0.07113394
#> [18,] 0.12052087 0.016562023 0.1143055 0.17096425 0.152835395 0.16696052
#> [19,] 0.35456950 0.032009918 0.3532110 0.29976484 0.272401698 0.25870535
#> [20,] 0.97270109 0.020145820 0.8906227 0.70976015 0.616189526 0.45898027
#> [21,] 0.16202811 0.012547036 0.1692129 0.17360383 0.284110337 0.22749928
#> [22,] 0.21343868 0.089516337 0.3615010 0.42502987 0.355328964 0.22726788
#> [23,] 0.29152862 0.005866361 0.1876999 0.13537527 0.241197849 0.24085260
#> [24,] 0.44227327 0.024597321 0.3163093 0.20074350 0.154707045 0.05229837
#> [25,] 0.20376794 0.061096094 0.2258863 0.13790007 0.006632596 0.06726861
#>              V44        V45        V46        V47        V48        V49
#>  [1,] 0.38039197 0.32089437 0.22699278 0.08388698 0.02290163 0.02776673
#>  [2,] 0.20797722 0.22448895 0.19081252 0.09080370 0.11435066 0.07010802
#>  [3,] 0.19160000 0.20850000 0.23350000 0.19640000 0.13000000 0.06330000
#>  [4,] 0.41244317 0.43247303 0.26630329 0.16340461 0.15506341 0.08825582
#>  [5,] 0.23300494 0.12688553 0.08280589 0.05153256 0.05385947 0.05368393
#>  [6,] 0.42035021 0.50561942 0.38719337 0.21237024 0.15553112 0.08597687
#>  [7,] 0.10044212 0.10882580 0.14789041 0.16425036 0.12000976 0.05286910
#>  [8,] 0.20001044 0.29047407 0.23220012 0.19901885 0.15625970 0.06842436
#>  [9,] 0.18791016 0.17374268 0.18238550 0.13456909 0.08261341 0.03877209
#> [10,] 0.19089141 0.13925936 0.10835810 0.10335123 0.08128598 0.04965903
#> [11,] 0.23544708 0.21358827 0.21545066 0.17092368 0.09946500 0.05573460
#> [12,] 0.01019626 0.17507518 0.14476741 0.05855643 0.06422935 0.05791225
#> [13,] 0.16940000 0.09540000 0.00800000 0.07900000 0.12550000 0.06470000
#> [14,] 0.50935445 0.57603028 0.67456540 0.48081224 0.28465565 0.14179987
#> [15,] 0.12155317 0.11012457 0.04190126 0.07270181 0.05699913 0.02107224
#> [16,] 0.15099094 0.14604985 0.09635146 0.12351292 0.08459636 0.04922246
#> [17,] 0.13878708 0.08712544 0.07257643 0.09288921 0.06488043 0.04974150
#> [18,] 0.13655437 0.08393153 0.06681775 0.03972020 0.03312071 0.01788908
#> [19,] 0.23160819 0.16360798 0.10390944 0.07771668 0.06632576 0.03337563
#> [20,] 0.38407217 0.39271579 0.35561968 0.27506533 0.17597051 0.09251772
#> [21,] 0.16563556 0.12052736 0.02753787 0.05091944 0.06473603 0.01585568
#> [22,] 0.18857939 0.19306434 0.13519336 0.14579546 0.13566431 0.08757746
#> [23,] 0.16215540 0.14257657 0.11892871 0.07533383 0.05605680 0.02334400
#> [24,] 0.04851982 0.08371705 0.07107641 0.06359277 0.04679964 0.02045841
#> [25,] 0.09787207 0.05176372 0.02031944 0.06499841 0.02519666 0.02341985
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.04250475 0.009098042 0.001746642 0.010892162 0.007630299 0.005405875
#>  [2,] 0.06008918 0.016488740 0.008174685 0.020445075 0.018709406 0.016197846
#>  [3,] 0.07210000 0.018300000 0.013700000 0.015000000 0.007600000 0.003200000
#>  [4,] 0.09835922 0.027930220 0.024609424 0.014396662 0.013473254 0.012860407
#>  [5,] 0.11111547 0.018994180 0.022258732 0.013251966 0.013965283 0.023699847
#>  [6,] 0.07235237 0.034771644 0.021337350 0.021219984 0.013758354 0.016341800
#>  [7,] 0.07377688 0.012880720 0.017415386 0.012627895 0.008229425 0.007085931
#>  [8,] 0.41612945 0.019950301 0.015600000 0.037664283 0.021690115 0.015616767
#>  [9,] 0.03243031 0.008021096 0.022424030 0.014775655 0.008976843 0.007466294
#> [10,] 0.10600076 0.020227991 0.016810046 0.013157257 0.014238436 0.015071444
#> [11,] 0.06627151 0.021469426 0.017782463 0.011822377 0.005146232 0.005459520
#> [12,] 0.22954650 0.054288693 0.025300739 0.040280429 0.025186727 0.017547728
#> [13,] 0.07600000 0.017900000 0.005100000 0.006100000 0.009300000 0.013500000
#> [14,] 0.06868951 0.050517589 0.085688044 0.059145346 0.027515896 0.028027568
#> [15,] 0.07597950 0.011231043 0.009334466 0.014175430 0.011902770 0.004382241
#> [16,] 0.04756468 0.021591700 0.002026202 0.006679122 0.003638393 0.004882927
#> [17,] 0.03454613 0.030954535 0.014663207 0.009772186 0.009485905 0.002162078
#> [18,] 0.04255700 0.009663435 0.008905458 0.009024719 0.005965800 0.008832917
#> [19,] 0.04970855 0.013395530 0.012429890 0.007423551 0.013093303 0.013521836
#> [20,] 0.03950334 0.036027681 0.023803767 0.013624606 0.010567392 0.010732364
#> [21,] 0.03575877 0.005486638 0.001074896 0.002774237 0.008522032 0.009481366
#> [22,] 0.08304179 0.022720300 0.019017112 0.014082524 0.008858433 0.012680057
#> [23,] 0.03796718 0.026003888 0.011817456 0.001883705 0.001598836 0.001106819
#> [24,] 0.01066078 0.017291676 0.016655695 0.011929155 0.019065055 0.024358739
#> [25,] 0.06933468 0.024490943 0.019934592 0.007322063 0.015010706 0.015135455
#>               V55         V56           V57          V58          V59
#>  [1,] 0.002697858 0.001465198 -0.0001329723 0.0003876549 0.0024106410
#>  [2,] 0.031481815 0.009022022  0.0257561593 0.0242417799 0.0198383350
#>  [3,] 0.003700000 0.007100000  0.0040000000 0.0009000000 0.0015000000
#>  [4,] 0.016006538 0.009465752  0.0099255624 0.0102788151 0.0087659652
#>  [5,] 0.013160678 0.011926740  0.0072760845 0.0128734557 0.0156774754
#>  [6,] 0.016639105 0.011973318  0.0104778598 0.0105507575 0.0116840344
#>  [7,] 0.005302609 0.005873229  0.0042313992 0.0052078078 0.0041445908
#>  [8,] 0.018309667 0.001291152  0.0107504098 0.0130494815 0.0181140910
#>  [9,] 0.003202934 0.003613482  0.0033340203 0.0049348934 0.0045739086
#> [10,] 0.008271369 0.010403198  0.0068057642 0.0103721686 0.0072664130
#> [11,] 0.005015801 0.002859758  0.0033419517 0.0022318263 0.0031353837
#> [12,] 0.026715227 0.005318197  0.0016441570 0.0150726040 0.0211388579
#> [13,] 0.006300000 0.006300000  0.0034000000 0.0032000000 0.0062000000
#> [14,] 0.021303945 0.011238028  0.0137632655 0.0195431036 0.0182885778
#> [15,] 0.002921498 0.013723733  0.0106369147 0.0017602937 0.0063190507
#> [16,] 0.005861782 0.003781718  0.0040601813 0.0019036198 0.0006799404
#> [17,] 0.005355662 0.007182674  0.0052126487 0.0035577181 0.0039219903
#> [18,] 0.011948022 0.006846282  0.0065206517 0.0045755879 0.0049260199
#> [19,] 0.007306749 0.008860881  0.0109309139 0.0082283388 0.0115906949
#> [20,] 0.009978201 0.008419933  0.0025421202 0.0107478251 0.0077497832
#> [21,] 0.006944543 0.009860182  0.0041165390 0.0035034913 0.0032571344
#> [22,] 0.008391542 0.007175345  0.0099305125 0.0089326750 0.0101901149
#> [23,] 0.001679044 0.004365097  0.0143901788 0.0146268903 0.0020667290
#> [24,] 0.015145321 0.011528614  0.0094592789 0.0141583961 0.0028159750
#> [25,] 0.010749683 0.005737648  0.0002289489 0.0020252325 0.0022974950
#>                 V6         V60         V7           V8          V9
#>  [1,]  0.066713097 0.004241174 0.07447118  0.009614410  0.07923859
#>  [2,]  0.080416677 0.009633398 0.11943398  0.279563674  0.10538010
#>  [3,]  0.134100000 0.008500000 0.16260000  0.190200000  0.26100000
#>  [4,]  0.108223093 0.004999489 0.12291311  0.161529709  0.25805209
#>  [5,]  0.115741490 0.009110165 0.14792726  0.157282158  0.11492405
#>  [6,]  0.103375006 0.010683292 0.15094304  0.247748020  0.38070682
#>  [7,]  0.104892946 0.003190436 0.11116482  0.109024409  0.15763133
#>  [8,]  0.184151998 0.023706063 0.18814894 -0.005170245  0.18566706
#>  [9,]  0.062523401 0.003302934 0.10449682  0.111150995  0.17837296
#> [10,]  0.138583301 0.005048081 0.13269873  0.117497598  0.16490980
#> [11,]  0.044753892 0.003916170 0.08726053  0.105207021  0.15410177
#> [12,]  0.191233075 0.011658423 0.15369322  0.061226922  0.11127250
#> [13,]  0.095800000 0.006700000 0.09900000  0.101800000  0.10300000
#> [14,]  0.142447351 0.010122593 0.12294756  0.190407232  0.17691333
#> [15,]  0.092846764 0.004855214 0.11580194  0.105079287  0.12337311
#> [16,]  0.022273368 0.001600903 0.08939126  0.075817657  0.11449932
#> [17,]  0.086040938 0.003229416 0.17292075  0.130461993  0.10778762
#> [18,]  0.081574084 0.003852343 0.09662647  0.106356298  0.12798920
#> [19,]  0.074859878 0.006642954 0.07869920  0.068782670  0.07416868
#> [20,]  0.061643860 0.010028018 0.12412547  0.154172376  0.08501012
#> [21,]  0.083933520 0.009153010 0.09972522  0.107443125  0.06044092
#> [22,]  0.109636584 0.007657314 0.14487867  0.188367963  0.23345113
#> [23,]  0.058198463 0.006752177 0.05099454  0.001159197  0.08711220
#> [24,] -0.003807518 0.003954455 0.02135436  0.041178710  0.09761437
#> [25,]  0.081734380 0.005730874 0.10262860  0.035547532 -0.02809893
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1884058