Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12       V13         V14
#>  [1,] 0.029841128 0.09001598 0.13749597 0.19635652 0.2837103  0.38155921
#>  [2,] 0.079862080 0.29517486 0.45068251 0.52533203 0.6101974  0.65504456
#>  [3,] 0.084404631 0.70273878 0.52038293 0.23808908 0.1343098  0.44169348
#>  [4,] 0.010758122 0.20741507 0.24737521 0.28737911 0.4277421  0.59310599
#>  [5,] 0.025739261 0.30869998 0.39331696 0.46998399 0.4676161  0.36171063
#>  [6,] 0.046840405 0.14711373 0.20312184 0.23241581 0.2920673  0.32945236
#>  [7,] 0.025735782 0.19353047 0.23265468 0.26970368 0.2830128  0.34419288
#>  [8,] 0.023529567 0.26146226 0.30942460 0.34436362 0.2952284  0.20576552
#>  [9,] 0.017612430 0.32245131 0.35796386 0.34555882 0.2563072  0.12330077
#> [10,] 0.024507601 0.37705721 0.40198621 0.41871390 0.4949812  0.42535215
#> [11,] 0.042420770 0.32800060 0.37324004 0.37098511 0.1851811 -0.01312340
#> [12,] 0.031972334 0.30783610 0.34424563 0.33705403 0.2798416  0.14774529
#> [13,] 0.059040991 0.52552958 0.37425610 0.12098455 0.2286986  0.32006085
#> [14,] 0.023951881 0.30802034 0.25121746 0.26889644 0.2385063  0.18355875
#> [15,] 0.027048433 0.12730008 0.14920109 0.19982017 0.2182056  0.18157107
#> [16,] 0.025462723 0.04583855 0.05259454 0.12260603 0.2145963  0.38622584
#> [17,] 0.039213850 0.24433317 0.27620795 0.24361569 0.2366792  0.27715086
#> [18,] 0.022247775 0.10754648 0.12790260 0.06968953 0.1091258  0.09287392
#> [19,] 0.028255483 0.12371327 0.11822191 0.17150800 0.1817752  0.15538667
#> [20,] 0.009668152 0.10421005 0.09886078 0.09552774 0.1110052  0.12160616
#> [21,] 0.004387991 0.19657161 0.11834131 0.12447717 0.2902858  0.39745008
#> [22,] 0.020403859 0.13111966 0.09291604 0.03189267 0.1458306  0.30133851
#> [23,] 0.016062747 0.06897751 0.20146259 0.38227730 0.2822607  0.24562782
#> [24,] 0.007663001 0.13727817 0.12098287 0.07895388 0.1299069  0.13967809
#> [25,] 0.026200000 0.61940000 0.63330000 0.70600000 0.5544000  0.53200000
#>              V15       V16       V17       V18       V19         V2       V20
#>  [1,] 0.50934691 0.5669692 0.6527561 0.7531641 0.9119560 0.03767754 1.0380096
#>  [2,] 0.78053058 0.8608701 0.8632486 0.9192868 0.9401550 0.10042690 0.9312268
#>  [3,] 0.47284121 0.4649971 0.2828545 0.3265699 0.4962589 0.07983520 0.9472739
#>  [4,] 0.59314268 0.5492336 0.6404072 0.7324608 0.8272136 0.03263428 0.8917803
#>  [5,] 0.24699648 0.2719408 0.3280980 0.3441107 0.4684059 0.02105087 0.6332866
#>  [6,] 0.31498484 0.2692739 0.3176920 0.1603472 0.1456487 0.05947158 0.3312912
#>  [7,] 0.40420899 0.4788596 0.5487876 0.6573383 0.7307957 0.02007732 0.8009363
#>  [8,] 0.15811096 0.2164736 0.2456665 0.3724577 0.5721332 0.03153266 0.6942407
#>  [9,] 0.11768165 0.2649063 0.3648638 0.4799558 0.5711519 0.02193557 0.6244217
#> [10,] 0.30692389 0.3320347 0.3425322 0.3013472 0.4863863 0.03798018 0.6922690
#> [11,] 0.02795284 0.1489847 0.1726680 0.2090254 0.2321882 0.04104910 0.4157375
#> [12,] 0.16311100 0.1977257 0.2056047 0.2416348 0.2104201 0.04433354 0.1642086
#> [13,] 0.42792491 0.4426223 0.1863866 0.1588195 0.3060187 0.08212388 0.8818770
#> [14,] 0.12315246 0.1784012 0.1729994 0.2188581 0.2038570 0.02991483 0.1722852
#> [15,] 0.20488470 0.3020713 0.3574984 0.3654600 0.3942989 0.03316270 0.3703110
#> [16,] 0.50832618 0.6707037 0.7477346 0.7778677 0.7771006 0.03381738 0.8544642
#> [17,] 0.33631815 0.3486273 0.3867396 0.2914349 0.4290773 0.03055547 0.6402031
#> [18,] 0.10316887 0.1519889 0.1811867 0.1977072 0.2032820 0.01341770 0.2782849
#> [19,] 0.14746235 0.2486558 0.2744645 0.3963166 0.3840823 0.02134134 0.3545046
#> [20,] 0.17498586 0.1749959 0.1435431 0.1515495 0.1043331 0.01336363 0.1318997
#> [21,] 0.53134524 0.4161737 0.1166706 0.1548845 0.2398952 0.03261621 0.2994221
#> [22,] 0.55218313 0.7866405 0.8410486 0.8056276 0.8765378 0.01500772 0.8795511
#> [23,] 0.11475858 0.1022814 0.1838762 0.4222689 0.3024221 0.05337336 0.6030323
#> [24,] 0.17071170 0.2316561 0.2833007 0.3593400 0.4324116 0.01078959 0.4434481
#> [25,] 0.64790000 0.6931000 0.6759000 0.7551000 0.8929000 0.05820000 0.8619000
#>             V21       V22        V23        V24       V25       V26       V27
#>  [1,] 1.0446068 1.0624865 1.01346722 0.81823697 0.5744893 0.5117258 0.4990929
#>  [2,] 0.8484400 0.6605339 0.57559996 0.58837087 0.4780362 0.3556197 0.3425114
#>  [3,] 1.0601438 0.4032563 0.09374677 0.02695704 0.2355713 0.4338352 0.2365337
#>  [4,] 0.9881941 1.0009470 0.89926455 0.87345649 0.8951881 0.8675850 0.7627233
#>  [5,] 0.7271784 0.8444027 0.86487595 0.87380544 0.9293866 0.9933823 0.9883132
#>  [6,] 0.4940370 0.5952485 0.62944525 0.66603715 0.6008786 0.5231944 0.6304593
#>  [7,] 0.8663484 0.9296504 0.94533609 0.97344376 0.9631141 0.9362723 0.8920750
#>  [8,] 0.8136800 0.7986202 0.83018068 0.93767135 0.9158325 0.9255153 0.9702217
#>  [9,] 0.7011551 0.8161815 0.93723228 0.98792841 0.9162625 0.7604467 0.5177053
#> [10,] 0.7103413 0.6726945 0.63137378 0.62656600 0.6895100 0.8324956 0.9027609
#> [11,] 0.5544567 0.6943956 0.72646574 0.74272739 0.9195645 1.0598074 1.0339878
#> [12,] 0.3044342 0.5825629 0.77564649 0.83372968 0.8427981 0.8503398 0.8906815
#> [13,] 0.9562546 0.3731777 0.14795226 0.18995676 0.1402552 0.3298594 0.2570883
#> [14,] 0.1144728 0.2020581 0.37781393 0.50235600 0.6013916 0.7122155 0.8092037
#> [15,] 0.3116221 0.3021374 0.34547585 0.39689469 0.4467337 0.5232190 0.5822320
#> [16,] 1.0000000 0.9084745 0.84087083 0.89492429 0.9077701 0.9352593 0.8640106
#> [17,] 0.5517352 0.5160108 0.83344308 0.79288660 0.5186831 0.5362313 0.6900032
#> [18,] 0.4069194 0.6235535 0.80183725 0.82682913 0.8110961 0.8537105 0.8194683
#> [19,] 0.5460693 0.6120154 0.67905482 0.72423331 0.5736588 0.5195097 0.4682387
#> [20,] 0.2127671 0.2395107 0.36700838 0.59133328 0.8360080 0.9864998 0.8456836
#> [21,] 0.2784684 0.3169507 0.50219553 0.76149429 0.7220994 0.7164875 0.6463644
#> [22,] 0.9518123 0.9551700 0.95570016 0.96331114 0.9164455 0.9612606 0.9655225
#> [23,] 0.7269683 0.7549340 0.79558862 0.84783504 1.0271503 0.9190356 0.6848965
#> [24,] 0.4798537 0.5332899 0.56549455 0.66494774 0.8417214 0.8400501 0.8082489
#> [25,] 0.7974000 0.6737000 0.42930000 0.36480000 0.5331000 0.2413000 0.5070000
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.5314386 0.5240283 0.066964280 0.4430026 0.3581435 0.3304341 0.33449153
#>  [2,] 0.3755577 0.3649770 0.099256054 0.2677692 0.2738983 0.2542431 0.17479522
#>  [3,] 0.5432742 0.6696101 0.037437265 0.8431126 0.4097155 0.3496450 0.55591132
#>  [4,] 0.6369309 0.5515349 0.027237766 0.3854531 0.2324024 0.1061681 0.09268732
#>  [5,] 0.7959424 0.5563147 0.023358404 0.4267719 0.3439752 0.4114311 0.38750709
#>  [6,] 0.7239494 0.7700627 0.116140405 0.7923513 0.8430514 0.8686306 0.70150540
#>  [7,] 0.8043403 0.7027419 0.027049915 0.5687011 0.4239461 0.3126887 0.21922835
#>  [8,] 0.9930977 0.8620671 0.037948302 0.6096249 0.3906498 0.2799111 0.20237244
#>  [9,] 0.3341676 0.2123886 0.027496441 0.3923790 0.5370825 0.5168659 0.64167419
#> [10,] 0.8871430 0.7732239 0.052858584 0.6519549 0.6150803 0.6618276 0.54684476
#> [11,] 0.8189024 0.7272768 0.059534545 0.7667881 0.7220121 0.4525604 0.26099257
#> [12,] 0.9633498 0.8221469 0.051387288 0.6109231 0.6606374 0.6380581 0.48260188
#> [13,] 0.4879709 0.4609187 0.087335704 0.6757096 0.2465040 0.3195846 0.60290579
#> [14,] 0.8823193 0.9075598 0.026203028 0.9535543 0.9118983 0.8020204 0.68338253
#> [15,] 0.6837825 0.7583776 0.039802159 0.7834651 0.7558361 0.7015238 0.70720183
#> [16,] 0.6359812 0.3083411 0.013035949 0.2830739 0.3361260 0.2538149 0.24057122
#> [17,] 0.7848761 0.5742253 0.052430997 0.3765029 0.3050412 0.2512856 0.15997932
#> [18,] 0.8412358 0.9126614 0.020784767 0.8230488 0.7061258 0.6419955 0.67271142
#> [19,] 0.3137802 0.3853477 0.033222450 0.4787584 0.5572900 0.5672314 0.65980153
#> [20,] 0.6220286 0.5066410 0.011928188 0.3888163 0.3037641 0.1667813 0.17745695
#> [21,] 0.8022210 0.9521223 0.049363876 0.8815012 0.8062075 0.6637455 0.55108609
#> [22,] 0.7851518 0.4901275 0.036342960 0.3864191 0.4233011 0.3938658 0.32074679
#> [23,] 0.5139288 0.4685911 0.066963120 0.5021374 0.5926263 0.5434557 0.63667802
#> [24,] 0.7071213 0.4990394 0.006331834 0.3187138 0.2677998 0.3266914 0.30084370
#> [25,] 0.8533000 0.6036000 0.109900000 0.8514000 0.8512000 0.5045000 0.18620000
#>              V34        V35         V36        V37        V38          V39
#>  [1,] 0.29654600 0.26911224  0.26432315 0.28585155 0.39092527  0.310848941
#>  [2,] 0.25624701 0.33298576  0.16985088 0.15503911 0.17276735  0.205670923
#>  [3,] 0.79255958 0.80930014  0.72283449 0.70320176 0.80740047  0.748752864
#>  [4,] 0.05124172 0.07162395  0.18950158 0.19206536 0.30891734  0.410751112
#>  [5,] 0.34394340 0.21263963  0.39025236 0.46500090 0.38544856  0.349320372
#>  [6,] 0.58725264 0.66073150  0.43352917 0.15754196 0.34087525  0.326956157
#>  [7,] 0.12600016 0.14678701  0.13463307 0.17916157 0.25990052  0.280681694
#>  [8,] 0.15996519 0.13079356  0.02094429 0.04336970 0.09192225  0.096480348
#>  [9,] 0.69791950 0.58228205  0.39401708 0.22315238 0.21753091  0.301064135
#> [10,] 0.34645796 0.10938953  0.22293665 0.24350402 0.34566521  0.547472321
#> [11,] 0.14516270 0.09391897 -0.06444155 0.10359915 0.14616972  0.075048366
#> [12,] 0.23308166 0.11643446  0.07321943 0.16619613 0.16887931  0.166748428
#> [13,] 0.78468678 0.84985058  0.88562764 0.82195729 0.90815038  0.723718697
#> [14,] 0.60442533 0.60279558  0.63621982 0.66899500 0.59304882  0.547817567
#> [15,] 0.71326652 0.74138898  0.76066965 0.67300732 0.53922974  0.417358720
#> [16,] 0.18255780 0.11956400  0.03963292 0.10997775 0.13237203  0.122388987
#> [17,] 0.26587911 0.30335722  0.34187023 0.33605931 0.27110983  0.257815686
#> [18,] 0.70898648 0.67255139  0.53902582 0.37126425 0.26539661  0.231243903
#> [19,] 0.56137077 0.52359435  0.56290610 0.63663540 0.65048486  0.575963545
#> [20,] 0.21406013 0.27051615  0.19884999 0.07442961 0.05644873  0.135632965
#> [21,] 0.49048431 0.43913603  0.41312251 0.20330436 0.18799179  0.224537390
#> [22,] 0.29225439 0.23450364  0.11876954 0.25475170 0.19144307  0.085225391
#> [23,] 0.55917650 0.55231806  0.54336638 0.44017967 0.21483530 -0.007770657
#> [24,] 0.25539447 0.22966593  0.24965729 0.28123273 0.27042025  0.318838170
#> [25,] 0.27090000 0.42320000  0.30430000 0.61160000 0.67560000  0.537500000
#>               V4        V40        V41        V42        V43       V44
#>  [1,] 0.09097429 0.19129990 0.20291092 0.34727802 0.34644600 0.2334447
#>  [2,] 0.07008952 0.16906786 0.09009871 0.10305649 0.17075054 0.1759410
#>  [3,] 0.17465031 0.57648438 0.41917461 0.45637101 0.44768770 0.3834458
#>  [4,] 0.04306869 0.30067143 0.19327992 0.15874308 0.12470178 0.1679966
#>  [5,] 0.02204320 0.28664153 0.13519113 0.17702114 0.34133599 0.4160000
#>  [6,] 0.11590606 0.19771069 0.58278966 0.58347872 0.29992012 0.5072973
#>  [7,] 0.06156617 0.29310682 0.28273839 0.27911185 0.22830499 0.1789512
#>  [8,] 0.04348840 0.11219807 0.05260538 0.08154744 0.18366559 0.2188108
#>  [9,] 0.03109365 0.40894798 0.34209876 0.32097275 0.29123069 0.1795473
#> [10,] 0.04828539 0.46439553 0.33679996 0.44016680 0.44326406 0.4676749
#> [11,] 0.07524845 0.04199726 0.03925174 0.11828443 0.16192710 0.1581368
#> [12,] 0.07525240 0.12682558 0.20890593 0.09751930 0.10010978 0.1437681
#> [13,] 0.08833090 0.49530284 0.38580799 0.25983714 0.15736693 0.2029305
#> [14,] 0.04463689 0.56735236 0.50525546 0.43275107 0.44282889 0.5139331
#> [15,] 0.03725010 0.38867497 0.34611495 0.30504010 0.23999120 0.2201801
#> [16,] 0.03568800 0.15343326 0.11485282 0.05207826 0.06755918 0.1355573
#> [17,] 0.04569913 0.16278617 0.12052909 0.12650349 0.12042823 0.1063952
#> [18,] 0.03321619 0.27451009 0.22729004 0.18118830 0.22270211 0.1657598
#> [19,] 0.03553522 0.65823879 0.66584138 0.58079357 0.53462566 0.4721673
#> [20,] 0.01609491 0.20424411 0.27224830 0.26473453 0.20370648 0.1232532
#> [21,] 0.02565328 0.12256579 0.11462725 0.10450505 0.18546928 0.1503954
#> [22,] 0.03810300 0.09773735 0.06915092 0.03650390 0.08260919 0.1206907
#> [23,] 0.05954256 0.29843026 0.36647481 0.55515229 0.44966481 0.2250536
#> [24,] 0.01167875 0.27266352 0.22285232 0.18585441 0.11992296 0.1100720
#> [25,] 0.10830000 0.47190000 0.46470000 0.25870000 0.21290000 0.2222000
#>               V45        V46        V47         V48          V49         V5
#>  [1,] 0.144819074 0.12265043 0.07737409 0.070141519  0.053399232 0.16822763
#>  [2,] 0.087301668 0.04576363 0.07609156 0.078596042  0.055161138 0.05819423
#>  [3,] 0.517392683 0.41941237 0.07317412 0.140668980  0.073169544 0.14200219
#>  [4,] 0.140673267 0.10843052 0.11920891 0.100954066  0.075762905 0.12717685
#>  [5,] 0.367621449 0.23350981 0.11009434 0.040718409  0.019440339 0.04425618
#>  [6,] 0.605030033 0.51661335 0.33601348 0.144419504  0.058456645 0.03865646
#>  [7,] 0.132926521 0.10373191 0.10352888 0.072561924  0.051364341 0.10638030
#>  [8,] 0.189615328 0.20592120 0.17203434 0.105393651  0.051335976 0.06004673
#>  [9,] 0.139942988 0.10574103 0.11704880 0.118925235  0.065967179 0.03523554
#> [10,] 0.539564217 0.31307466 0.20187775 0.204166925  0.112019595 0.09268357
#> [11,] 0.131323396 0.09477404 0.12744476 0.104090431  0.060061226 0.01881035
#> [12,] 0.089777300 0.08519834 0.11277774 0.138319186  0.080106300 0.08494261
#> [13,] 0.367375559 0.26834520 0.17485492 0.194784311  0.099716354 0.04531914
#> [14,] 0.510391695 0.38523177 0.28604018 0.222728124  0.120939767 0.06767756
#> [15,] 0.171310988 0.14489427 0.09740791 0.074431650  0.036212218 0.04539410
#> [16,] 0.089715626 0.06204718 0.08676272 0.061345175  0.044344493 0.04413847
#> [17,] 0.101824441 0.08873537 0.05951477 0.024072696  0.024788070 0.06210171
#> [18,] 0.100437628 0.08585254 0.05815245 0.042597649  0.015004440 0.05356144
#> [19,] 0.337217404 0.25456159 0.16267062 0.103729661  0.032216000 0.07596011
#> [20,] 0.098875290 0.09981066 0.04924322 0.037864436  0.030373490 0.03082418
#> [21,] 0.001646884 0.02496522 0.03203923 0.006427182 -0.007597318 0.04995731
#> [22,] 0.110875887 0.03645356 0.04782285 0.041929872  0.020670826 0.06967708
#> [23,] 0.148356678 0.12103924 0.08417164 0.086648040  0.033405492 0.13013199
#> [24,] 0.092240712 0.09157757 0.03953355 0.016315121  0.012150357 0.03079179
#> [25,] 0.211100000 0.01760000 0.13480000 0.074400000  0.013000000 0.09740000
#>               V50         V51         V52         V53         V54         V55
#>  [1,] 0.016092455 0.014803192 0.022478399 0.014522825 0.020893860 0.009030759
#>  [2,] 0.015843322 0.029622623 0.020636010 0.018271133 0.018899455 0.018233796
#>  [3,] 0.045282908 0.019235713 0.033658213 0.038855602 0.029354791 0.018366081
#>  [4,] 0.029429971 0.023227067 0.013940821 0.012007707 0.022323319 0.020060412
#>  [5,] 0.009544934 0.007082063 0.005600136 0.008756941 0.009016284 0.005274421
#>  [6,] 0.028623851 0.037035839 0.036944720 0.018147701 0.029364255 0.030267299
#>  [7,] 0.023431535 0.019053542 0.015609864 0.017228887 0.012446406 0.006457659
#>  [8,] 0.012907351 0.017761091 0.012645711 0.005242623 0.005395373 0.004249731
#>  [9,] 0.018050008 0.012971486 0.009501358 0.008926202 0.008743431 0.006725554
#> [10,] 0.033564271 0.027486715 0.017034017 0.010558899 0.011638632 0.015667418
#> [11,] 0.022489854 0.003548377 0.009720056 0.002651535 0.011778841 0.004291023
#> [12,] 0.023575877 0.018611322 0.015035517 0.008225236 0.011994061 0.007031743
#> [13,] 0.031287440 0.034047514 0.017259386 0.012817741 0.027376510 0.018161354
#> [14,] 0.033967293 0.019925415 0.018632195 0.008807073 0.007287694 0.005710498
#> [15,] 0.015574924 0.015226800 0.011379924 0.012434900 0.010709838 0.008212448
#> [16,] 0.028082196 0.012677242 0.011590094 0.010015030 0.002425021 0.004817464
#> [17,] 0.013813320 0.020960523 0.022776986 0.011134075 0.004128699 0.009857940
#> [18,] 0.009814023 0.005208028 0.005748983 0.007056022 0.007544342 0.006702137
#> [19,] 0.016601695 0.016046308 0.008035070 0.008661858 0.007301070 0.002092676
#> [20,] 0.009994597 0.004828840 0.009539879 0.003364113 0.003326753 0.005123638
#> [21,] 0.023126239 0.017092344 0.017183361 0.016838500 0.005355846 0.005753614
#> [22,] 0.007719272 0.015638484 0.009668005 0.010893838 0.008436191 0.003096304
#> [23,] 0.007136914 0.006671793 0.003060950 0.011010978 0.008573339 0.013024376
#> [24,] 0.013348484 0.003407738 0.005219542 0.007683873 0.011239108 0.004834026
#> [25,] 0.010600000 0.003300000 0.023200000 0.016600000 0.009500000 0.018000000
#>               V56         V57         V58          V59         V6          V60
#>  [1,] 0.012604107 0.009607488 0.015113981 0.0114174849 0.16705401 0.0005860087
#>  [2,] 0.011484536 0.005858694 0.016574661 0.0131394845 0.11416739 0.0140245031
#>  [3,] 0.026287054 0.014895137 0.008078668 0.0036162603 0.03671773 0.0114995180
#>  [4,] 0.006162127 0.008497678 0.008937696 0.0094564073 0.18099250 0.0058526755
#>  [5,] 0.006502808 0.003037563 0.003695953 0.0072504764 0.09098902 0.0046365211
#>  [6,] 0.014152299 0.015001459 0.010985062 0.0300822059 0.09222854 0.0159030753
#>  [7,] 0.010152702 0.008007064 0.005924912 0.0073051624 0.13332140 0.0049501433
#>  [8,] 0.004019584 0.003635255 0.002834945 0.0016300023 0.09714170 0.0040085496
#>  [9,] 0.005944862 0.002969771 0.007176765 0.0039401577 0.07321841 0.0034017679
#> [10,] 0.008685155 0.009512071 0.010340044 0.0072132379 0.12179637 0.0047672830
#> [11,] 0.002727284 0.003593063 0.009044115 0.0032053081 0.15999614 0.0043061025
#> [12,] 0.010406665 0.007171541 0.008004318 0.0103297510 0.13782870 0.0096510105
#> [13,] 0.022225836 0.016977832 0.016158006 0.0180725091 0.12172442 0.0120056581
#> [14,] 0.008496011 0.007582445 0.004690663 0.0040664871 0.08263752 0.0027536159
#> [15,] 0.007041248 0.008430835 0.010470907 0.0113376986 0.08998883 0.0066327407
#> [16,] 0.009382706 0.007422672 0.003282451 0.0048850553 0.08473988 0.0034725447
#> [17,] 0.007914013 0.008282634 0.011648053 0.0080065526 0.10251679 0.0065638800
#> [18,] 0.006761228 0.010212794 0.007742207 0.0049518070 0.07587797 0.0049778550
#> [19,] 0.009039204 0.006088740 0.006667914 0.0098487486 0.10347244 0.0084388919
#> [20,] 0.002524965 0.002706538 0.003294415 0.0007838652 0.05817926 0.0025887161
#> [21,] 0.007087802 0.004395249 0.002827167 0.0148334801 0.12756008 0.0084617495
#> [22,] 0.007472810 0.005222106 0.003848228 0.0048443796 0.09498521 0.0033955990
#> [23,] 0.008927095 0.009063037 0.007360330 0.0035309923 0.18150072 0.0150167429
#> [24,] 0.002876293 0.003483867 0.006999204 0.0037089389 0.01264231 0.0043072059
#> [25,] 0.024400000 0.031600000 0.016400000 0.0095000000 0.22800000 0.0078000000
#>               V7         V8         V9
#>  [1,] 0.09038288 0.10243743 0.12031479
#>  [2,] 0.15493749 0.13666894 0.20100544
#>  [3,] 0.09289528 0.30632461 0.61895390
#>  [4,] 0.20879841 0.19590080 0.20901186
#>  [5,] 0.10124531 0.09769700 0.25569304
#>  [6,] 0.04352258 0.05338938 0.21565249
#>  [7,] 0.15232110 0.13835863 0.15913413
#>  [8,] 0.13996653 0.15068722 0.19439295
#>  [9,] 0.09783292 0.14232920 0.24808315
#> [10,] 0.12336094 0.14882810 0.26685365
#> [11,] 0.16828315 0.15876471 0.25864996
#> [12,] 0.15839001 0.18187033 0.22113734
#> [13,] 0.23385072 0.41314027 0.58485830
#> [14,] 0.11497018 0.13766661 0.26233954
#> [15,] 0.10080003 0.08062803 0.10563708
#> [16,] 0.16025372 0.13233224 0.11017037
#> [17,] 0.12558717 0.17834360 0.20532420
#> [18,] 0.08397250 0.08733618 0.07676759
#> [19,] 0.13263677 0.11792529 0.12164073
#> [20,] 0.07541503 0.07201370 0.11127412
#> [21,] 0.13156657 0.16776007 0.20236961
#> [22,] 0.12530408 0.09006863 0.12835945
#> [23,] 0.14469942 0.16964183 0.04620889
#> [24,] 0.03872375 0.05453978 0.08651756
#> [25,] 0.24310000 0.37710000 0.55980000
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203