Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12         V13        V14
#>  [1,] 0.054655792 0.27146883 0.33104281 0.41497344  0.53201182 0.51865429
#>  [2,] 0.022955038 0.19342523 0.18520706 0.15942695  0.13744216 0.09638826
#>  [3,] 0.035706180 0.18794492 0.27305040 0.33481994  0.29820451 0.31434934
#>  [4,] 0.030350663 0.21716326 0.22040518 0.24697005  0.28928307 0.28685681
#>  [5,] 0.004402484 0.20262264 0.19507348 0.11624808 -0.01149109 0.09330144
#>  [6,] 0.018951143 0.16106321 0.31412785 0.47991567  0.55449451 0.55537710
#>  [7,] 0.027899935 0.30498877 0.35860589 0.39421814  0.37939226 0.32398487
#>  [8,] 0.103704130 0.35248329 0.42987720 0.43217678  0.65660306 0.87687271
#>  [9,] 0.085773424 0.12025947 0.20347474 0.31287287  0.32579924 0.22918183
#> [10,] 0.016866087 0.15487140 0.25559294 0.30303753  0.26993875 0.19772576
#> [11,] 0.020352674 0.24027896 0.30589457 0.33659088  0.32912013 0.27702503
#> [12,] 0.019734614 0.27033464 0.32304123 0.33355983  0.27212906 0.20065655
#> [13,] 0.033827167 0.23413531 0.21980970 0.17210799  0.19408048 0.37645866
#> [14,] 0.012914303 0.43281411 0.30916807 0.13422659  0.33508831 0.56980562
#> [15,] 0.085908375 0.27956802 0.27109518 0.08780661  0.09662056 0.45854492
#> [16,] 0.023950874 0.09656119 0.12271460 0.10594930  0.19961878 0.32577090
#> [17,] 0.013016247 0.10000586 0.13032882 0.12534601  0.18162933 0.19946951
#> [18,] 0.022556152 0.12616059 0.16014901 0.19531398  0.19603109 0.17622452
#> [19,] 0.012514588 0.16616757 0.13620538 0.17197962  0.29022159 0.13697438
#> [20,] 0.011417883 0.12507801 0.10782300 0.10265545  0.16338783 0.18988230
#> [21,] 0.022289716 0.21638412 0.23383498 0.13056245  0.19975942 0.46830564
#> [22,] 0.017329726 0.14992818 0.12888368 0.08056346  0.06503371 0.13307957
#> [23,] 0.013345113 0.07325947 0.04240883 0.09233646  0.11183640 0.09518976
#> [24,] 0.035943748 0.04000475 0.09031458 0.25483415  0.14121376 0.11381323
#> [25,] 0.014931593 0.15353507 0.11444910 0.12713368  0.15354430 0.13838000
#>              V15        V16        V17       V18       V19           V2
#>  [1,] 0.63254082 0.70426257 0.89332720 0.9230188 0.8917689  0.102476545
#>  [2,] 0.24751634 0.40876628 0.49647128 0.5124706 0.5956494  0.031211303
#>  [3,] 0.42841305 0.53635460 0.66786490 0.7125990 0.7512366  0.047979930
#>  [4,] 0.22964825 0.26979680 0.29685701 0.2221391 0.2364904  0.062443555
#>  [5,] 0.21187714 0.52971763 0.69320032 0.7797168 0.7885892  0.006408098
#>  [6,] 0.62665565 0.74589194 0.88530946 0.9843090 0.9744074  0.030422818
#>  [7,] 0.28001524 0.34163013 0.41833421 0.4997574 0.5808725  0.032002586
#>  [8,] 0.97331613 1.00887741 0.83334249 0.8450519 0.8976268  0.109435924
#>  [9,] 0.23297014 0.29429918 0.21924388 0.2639326 0.3774566  0.157572300
#> [10,] 0.05079712 0.27557516 0.43645693 0.6394897 0.8073615 -0.002681340
#> [11,] 0.19477865 0.14315238 0.17435888 0.2379836 0.3166413  0.027756194
#> [12,] 0.17060610 0.18313762 0.20217347 0.3388718 0.5714375  0.046993392
#> [13,] 0.48155931 0.49702975 0.52048856 0.4530565 0.4895998  0.046673533
#> [14,] 0.69508376 0.96858283 0.99250988 0.9252741 0.6272837  0.052045596
#> [15,] 0.68477410 0.67764593 0.71752791 0.5710921 0.3829132  0.042841627
#> [16,] 0.55245338 0.74850794 0.82849132 0.8225861 0.8646789  0.012989857
#> [17,] 0.23220681 0.30407149 0.27201123 0.2272618 0.1921127  0.011767337
#> [18,] 0.22103740 0.52569850 0.71765280 0.6280592 0.3898628  0.015775787
#> [19,] 0.21150563 0.25264238 0.24922002 0.2779569 0.2751012  0.047793427
#> [20,] 0.19110884 0.17047944 0.14583910 0.2024606 0.2877364  0.014092870
#> [21,] 0.65307588 0.74203563 0.80205061 0.6889803 0.4999200  0.034048238
#> [22,] 0.09048194 0.11119970 0.24190563 0.2659240 0.3308525  0.005461510
#> [23,] 0.11729668 0.08854553 0.08711435 0.1843461 0.1572372  0.035051672
#> [24,] 0.10756272 0.15538585 0.13296034 0.1167352 0.1781051  0.041163213
#> [25,] 0.08535199 0.15077712 0.25682460 0.3207260 0.3741234  0.032056047
#>              V20       V21          V22       V23       V24         V25
#>  [1,] 0.77226963 0.7503209  0.551807347 0.2979422 0.2375627  0.19866983
#>  [2,] 0.77537685 0.9201848  1.008331259 0.9008286 0.7312665  0.53279188
#>  [3,] 0.81299595 0.8708235  0.950080276 0.9929744 0.9900469  0.86296691
#>  [4,] 0.23772484 0.2691538  0.354385535 0.4830456 0.5826873  0.64615205
#>  [5,] 0.66065396 0.7039123  0.894877726 0.9370013 0.4694236  0.09823786
#>  [6,] 0.91987762 0.7790302  0.690313301 0.6518017 0.5754913  0.36045592
#>  [7,] 0.71446968 0.8198492  0.873719301 0.8413357 0.8658094  0.91811764
#>  [8,] 0.96042907 0.8843294  0.834199789 0.9859755 0.9187852  0.61395406
#>  [9,] 0.25618420 0.2745306  0.570245099 0.8692891 0.8845715  0.72967336
#> [10,] 0.77297761 0.8118019  0.861877261 0.9585996 0.9838113  1.06973533
#> [11,] 0.40694697 0.4828924  0.482950420 0.5207258 0.6257681  0.72540628
#> [12,] 0.69051810 0.7922003  0.806336006 0.7918142 0.9007722  0.93139008
#> [13,] 0.72961007 0.9721220  1.053721550 0.8272462 0.4921820  0.34235440
#> [14,] 0.46792047 0.5069708  0.507140343 0.7017545 0.8309717  0.80447990
#> [15,] 0.26651270 0.5295972  0.772895682 0.8647706 0.9362203  0.94530228
#> [16,] 0.84788479 0.9345778  0.950564369 0.8761838 0.8533059  0.85587695
#> [17,] 0.18792037 0.2444706  0.332380034 0.4717955 0.6225154  0.78201026
#> [18,] 0.34554886 0.4441607  0.542993591 0.5924298 0.6604349  0.82924368
#> [19,] 0.09082225 0.2029807  0.185892232 0.1997025 0.2445593  0.29222843
#> [20,] 0.30441963 0.3069306  0.393281145 0.5146689 0.6428624  0.65952808
#> [21,] 0.32941786 0.4927154  0.721353277 0.8300423 0.9519390  0.80104191
#> [22,] 0.48075416 0.6033893  0.781188015 0.9234014 0.9161100  0.87913199
#> [23,] 0.22642865 0.2537234  0.163384091 0.1093420 0.1830816  0.16840040
#> [24,] 0.10947189 0.1985811 -0.003265454 0.2321823 0.2347695 -0.01072345
#> [25,] 0.47426003 0.5761908  0.605012866 0.6988128 0.7707486  0.89401804
#>             V26        V27        V28       V29           V3       V30
#>  [1,] 0.2015793 0.22827342 0.22121052 0.1514841  0.105844163 0.1493295
#>  [2,] 0.2780751 0.08753151 0.04742458 0.2412815  0.011070390 0.3815140
#>  [3,] 0.7167032 0.59768262 0.41483598 0.3685887  0.071929981 0.3028967
#>  [4,] 0.6900350 0.73242339 0.87280011 0.8875674  0.072267488 0.8768658
#>  [5,] 0.2778273 0.62212770 0.85756306 0.9821785 -0.019144955 0.8439277
#>  [6,] 0.1039015 0.05773849 0.07636775 0.1033929  0.033618599 0.1888935
#>  [7,] 0.9578998 0.95859840 0.89308322 0.7578399  0.033408619 0.5518414
#>  [8,] 0.5918904 0.57307081 0.57584720 0.5267063  0.065108379 0.3360456
#>  [9,] 0.5888934 0.53084323 0.58986160 0.6797365  0.194744629 0.5078331
#> [10,] 0.9141285 0.53140336 0.39086410 0.3677323  0.006336212 0.5518923
#> [11,] 0.8473079 0.94140977 0.97830070 0.9190464  0.037630163 0.7307893
#> [12,] 0.9978423 1.01466262 0.96730892 0.8678592  0.037776552 0.6752346
#> [13,] 0.5385963 0.60245364 0.36275492 0.2444194  0.036528136 0.3984134
#> [14,] 0.7568046 0.80348596 0.62568475 0.2441047  0.063204683 0.2731791
#> [15,] 0.6249421 0.59603833 0.67742513 0.5073909  0.036532489 0.5791538
#> [16,] 0.9083673 0.86001292 0.67819805 0.5011031  0.027549005 0.4838845
#> [17,] 0.8856855 0.81282775 0.74234989 0.6451560  0.019243309 0.5200557
#> [18,] 0.9551911 0.95858981 0.90220827 0.7966611  0.019850355 0.8025627
#> [19,] 0.3093407 0.15290205 0.50715747 0.7052426  0.031369137 0.6687246
#> [20,] 0.6723697 0.66898390 0.80412116 0.9660109  0.023782223 0.9663696
#> [21,] 0.6645493 0.67650317 0.54886865 0.4439716  0.082114739 0.4968911
#> [22,] 0.9393205 0.94052809 0.88623673 0.8524513  0.012869076 0.7501499
#> [23,] 0.2535558 0.31463672 0.47936664 0.6100749  0.024206645 0.7022109
#> [24,] 0.1745345 0.47121366 0.71199515 0.7028134  0.029785616 0.6650937
#> [25,] 0.8919151 0.83120826 0.63606961 0.5404628  0.042194063 0.4737010
#>             V31       V32        V33        V34         V35        V36
#>  [1,] 0.2227691 0.2854350 0.35387663 0.40820636  0.50675903 0.16826662
#>  [2,] 0.3480758 0.5542731 0.43042559 0.15980614 -0.02356838 0.21700637
#>  [3,] 0.2648378 0.2246880 0.36841639 0.48037992  0.50979407 0.34392414
#>  [4,] 0.9382331 0.8937755 0.70197159 0.54215362  0.53723735 0.44961637
#>  [5,] 0.4895977 0.5017827 0.46322988 0.41673666  0.28122698 0.17471079
#>  [6,] 0.3155852 0.4292867 0.37466097 0.48503999  0.46104420 0.34466802
#>  [7,] 0.4018517 0.3811833 0.30605371 0.19842328  0.11801644 0.18927433
#>  [8,] 0.2517090 0.1899085 0.06705201 0.12596270  0.15794063 0.16190689
#>  [9,] 0.4897607 0.3746712 0.29719673 0.20746741  0.14567134 0.46908976
#> [10,] 0.7332593 0.8844009 0.99181487 0.90977966  0.68349693 0.44577613
#> [11,] 0.6327062 0.5591239 0.52235528 0.40744817  0.22260599 0.12163050
#> [12,] 0.4684312 0.3246242 0.25235319 0.22199903  0.17889458 0.06351589
#> [13,] 0.2021839 0.2712340 0.34038386 0.24146317  0.15102148 0.25003195
#> [14,] 0.6350144 0.6722319 0.28789420 0.43540819  0.84296941 0.70843918
#> [15,] 0.4737987 0.2357947 0.28074391 0.19295457  0.52666701 0.74158528
#> [16,] 0.4875036 0.3668989 0.32638254 0.31354097  0.19184245 0.08620055
#> [17,] 0.4470780 0.3554964 0.31141896 0.32542875  0.38967734 0.33502307
#> [18,] 0.8156183 0.7128616 0.68397397 0.68975561  0.75215572 0.74581081
#> [19,] 0.5305878 0.5693404 0.65704099 0.60438494  0.58999161 0.74879665
#> [20,] 0.8785024 0.7781509 0.75569305 0.66299172  0.56883869 0.48996500
#> [21,] 0.3965679 0.1871448 0.19058585 0.08268089  0.39562509 0.66966905
#> [22,] 0.6677321 0.6174853 0.58618066 0.62305606  0.63278138 0.53542357
#> [23,] 0.7016703 0.6526211 0.73668561 0.88818051  0.94905746 0.90999995
#> [24,] 0.7269893 0.5424097 0.66720759 0.89155193  0.97381890 0.97471518
#> [25,] 0.4365041 0.5243834 0.55832417 0.49790229  0.42766642 0.41231301
#>              V37        V38       V39           V4       V40        V41
#>  [1,] 0.12575266 0.28006200 0.3036339  0.155101826 0.1658173 0.06408891
#>  [2,] 0.38704826 0.38217180 0.2911267  0.046609627 0.1814150 0.16551505
#>  [3,] 0.29595614 0.22088858 0.3179030  0.063092466 0.4274808 0.36058528
#>  [4,] 0.39041073 0.38884909 0.3769502  0.078656487 0.3142532 0.51229144
#>  [5,] 0.20214741 0.49828867 0.5072516 -0.001657471 0.3634116 0.07174836
#>  [6,] 0.27534908 0.40944582 0.3323854  0.041161409 0.1864044 0.17803783
#>  [7,] 0.24198538 0.34071974 0.3570795  0.055173084 0.3232947 0.22144048
#>  [8,] 0.13412477 0.12602572 0.2065891  0.029414773 0.2540471 0.15542351
#>  [9,] 0.52464118 0.29283260 0.2791358  0.204717701 0.3099731 0.31476749
#> [10,] 0.07059891 0.07336049 0.3130672  0.036898875 0.5205713 0.54050105
#> [11,] 0.19247069 0.19510098 0.2619128  0.046623168 0.2826711 0.31651134
#> [12,] 0.07558740 0.08797656 0.1058285  0.039199804 0.1206164 0.04001942
#> [13,] 0.28866287 0.29730463 0.2436188  0.054326375 0.1797424 0.13976170
#> [14,] 0.48313903 0.58767995 0.5001828  0.053105744 0.2908824 0.06945052
#> [15,] 0.78236091 0.44138122 0.1322368  0.018285281 0.3486871 0.44644569
#> [16,] 0.20579246 0.14917902 0.0919209  0.038366543 0.1436413 0.08273833
#> [17,] 0.12885903 0.06427480 0.1551964  0.025118504 0.2179506 0.22134225
#> [18,] 0.70220539 0.62627856 0.4187266  0.027713385 0.2597710 0.22529007
#> [19,] 0.82710761 0.88847424 0.9567794  0.028683941 0.9245831 0.87888405
#> [20,] 0.37942105 0.25085567 0.1828423  0.026351686 0.1735635 0.18159654
#> [21,] 0.66425809 0.31724331 0.1049326  0.070538451 0.3227006 0.43163862
#> [22,] 0.45602308 0.35961063 0.2103978  0.021465203 0.1804726 0.17978437
#> [23,] 0.78222980 0.62063002 0.5450760  0.029018889 0.4552158 0.36327024
#> [24,] 0.77683177 0.68134899 0.6504234  0.022498714 0.6796436 0.60020726
#> [25,] 0.38631321 0.26980405 0.1876241  0.029653157 0.2563290 0.24432898
#>              V42        V43        V44         V45        V46         V47
#>  [1,] 0.09478469 0.09774115 0.03899520 0.070732878 0.08194114  0.09583677
#>  [2,] 0.08496693 0.04003709 0.07711888 0.042025404 0.03372642  0.03466697
#>  [3,] 0.37794445 0.33537214 0.27801948 0.204501895 0.09636848  0.10525946
#>  [4,] 0.54604016 0.46581955 0.53279229 0.607391021 0.48939494  0.30744456
#>  [5,] 0.17052204 0.50389575 0.40489198 0.086436960 0.20701195  0.16759720
#>  [6,] 0.19742578 0.16110073 0.05748925 0.062708250 0.04682346  0.03414115
#>  [7,] 0.27226327 0.30003981 0.29581122 0.233861831 0.11303322  0.09487827
#>  [8,] 0.13052139 0.18836587 0.25110640 0.121803256 0.06394858  0.07966139
#>  [9,] 0.43266028 0.30880925 0.09236190 0.189380366 0.25503197  0.29975996
#> [10,] 0.40188516 0.21865968 0.03910973 0.006461434 0.05829999  0.11004887
#> [11,] 0.35822280 0.28556220 0.19380965 0.185176499 0.15638226  0.16198425
#> [12,] 0.09740100 0.20122992 0.22375666 0.215044080 0.21648100  0.16283140
#> [13,] 0.24494992 0.22297446 0.12461513 0.098893210 0.07171567  0.09520258
#> [14,] 0.11998169 0.27229488 0.27983648 0.207876265 0.06488145  0.05108057
#> [15,] 0.41653528 0.30038148 0.22550101 0.257314978 0.27604467  0.23965457
#> [16,] 0.03402320 0.07621526 0.10706501 0.107008152 0.05092277  0.07845151
#> [17,] 0.20991168 0.21043187 0.16482681 0.097964563 0.09984241  0.05120077
#> [18,] 0.18864178 0.18953498 0.13971867 0.137264444 0.15031326  0.08185910
#> [19,] 0.78018600 0.59565749 0.36589971 0.467532375 0.53429831  0.41263688
#> [20,] 0.18042949 0.22695534 0.21868895 0.168401321 0.11173976  0.06071835
#> [21,] 0.37747355 0.22743239 0.18043298 0.202364230 0.16153989  0.16950941
#> [22,] 0.06600375 0.07809571 0.04132136 0.011826381 0.01861431  0.01983465
#> [23,] 0.32048320 0.19795507 0.16880239 0.142662042 0.09088771  0.08758663
#> [24,] 0.51799218 0.38065945 0.22413830 0.148997757 0.06870419  0.07354302
#> [25,] 0.24595579 0.14030663 0.08767379 0.030593790 0.01205875 -0.00442971
#>              V48         V49         V5         V50         V51         V52
#>  [1,] 0.10327629 0.073785541 0.14456029 0.034476209 0.026567648 0.036708397
#>  [2,] 0.03037406 0.009977355 0.04120111 0.008614008 0.012516322 0.007721157
#>  [3,] 0.08232423 0.041143616 0.10037701 0.002194421 0.001333638 0.008001551
#>  [4,] 0.18790106 0.103891303 0.08864208 0.032802608 0.023243342 0.024706936
#>  [5,] 0.08059456 0.072420108 0.01519384 0.027901072 0.011974872 0.020261455
#>  [6,] 0.02317278 0.019800457 0.06385490 0.009843994 0.012295969 0.011843350
#>  [7,] 0.09193164 0.060563829 0.08688765 0.017517833 0.020878388 0.013225490
#>  [8,] 0.09987345 0.090693841 0.06983227 0.016011766 0.029933208 0.016171444
#>  [9,] 0.25264844 0.137582455 0.18901842 0.027649651 0.032401479 0.044207014
#> [10,] 0.11898839 0.082877730 0.03373343 0.020452960 0.017730796 0.007005620
#> [11,] 0.14690475 0.068603983 0.08788068 0.012485661 0.020735122 0.013502046
#> [12,] 0.09379627 0.048903656 0.04563329 0.013548899 0.021775718 0.014477763
#> [13,] 0.07978837 0.055218169 0.06795106 0.031128696 0.016796553 0.012179303
#> [14,] 0.03542211 0.017102936 0.10697352 0.028281700 0.002713866 0.020855465
#> [15,] 0.20466066 0.132887817 0.03785812 0.035183249 0.012557868 0.016887056
#> [16,] 0.05445595 0.023945560 0.06562134 0.013430031 0.012560114 0.005500359
#> [17,] 0.03563333 0.025730698 0.04831402 0.011079101 0.011842630 0.010833853
#> [18,] 0.04435331 0.022786191 0.03235447 0.020710894 0.011196659 0.011459278
#> [19,] 0.27312042 0.181071562 0.07619779 0.068373611 0.036438573 0.016020812
#> [20,] 0.03828255 0.018438055 0.02458352 0.015037435 0.011156703 0.009754520
#> [21,] 0.15328607 0.111363796 0.04710504 0.027783853 0.019162074 0.010207759
#> [22,] 0.03157039 0.010999455 0.05227308 0.008881424 0.005678967 0.005071738
#> [23,] 0.07116033 0.032717408 0.01155120 0.014376247 0.016992973 0.005503876
#> [24,] 0.06855883 0.055235566 0.07537275 0.017061432 0.004338106 0.009169614
#> [25,] 0.01876900 0.007475138 0.06687694 0.008525992 0.003005912 0.002399939
#>                V53          V54          V55         V56           V57
#>  [1,]  0.015925290 0.0109678555 0.0179652771 0.010360444  0.0007226653
#>  [2,] -0.001175717 0.0040604960 0.0026765972 0.003046972  0.0082603397
#>  [3,]  0.017022302 0.0111728785 0.0022052718 0.012785016  0.0065023973
#>  [4,]  0.017496973 0.0130228731 0.0096134934 0.009433685  0.0102658474
#>  [5,]  0.010317406 0.0007617112 0.0008757816 0.004619387 -0.0036145108
#>  [6,]  0.010230154 0.0213975272 0.0212988985 0.008879925  0.0038225478
#>  [7,]  0.011452633 0.0110145641 0.0070519487 0.006454717  0.0053918087
#>  [8,]  0.016130031 0.0175023351 0.0253020877 0.003935640  0.0024620651
#>  [9,]  0.022452118 0.0250563254 0.0179558656 0.013491447  0.0157182763
#> [10,]  0.003405103 0.0002296057 0.0082729229 0.005766443  0.0006258394
#> [11,]  0.007386353 0.0074793557 0.0056270058 0.006215415  0.0039614880
#> [12,]  0.006949495 0.0058483189 0.0020054112 0.002764458  0.0036992040
#> [13,]  0.005358255 0.0123986819 0.0023229005 0.011022337  0.0027734854
#> [14,]  0.005072704 0.0196596506 0.0127247968 0.002713054  0.0043447835
#> [15,]  0.013825127 0.0078931472 0.0031426401 0.004949239  0.0261791875
#> [16,]  0.008951637 0.0068051004 0.0046501553 0.008358406  0.0047659909
#> [17,]  0.008985638 0.0066514757 0.0064888194 0.005315336  0.0079571801
#> [18,]  0.002963397 0.0043795413 0.0079977557 0.004895510  0.0108758842
#> [19,]  0.014967700 0.0111808948 0.0190232182 0.014884716  0.0028249946
#> [20,]  0.009784767 0.0083199610 0.0050116491 0.004756402  0.0045513651
#> [21,]  0.005290735 0.0103952569 0.0050820436 0.005402032  0.0098888017
#> [22,]  0.011709941 0.0090838748 0.0051644759 0.005881194  0.0055374421
#> [23,]  0.009136823 0.0075997394 0.0124122351 0.009962819  0.0056509418
#> [24,]  0.016215803 0.0159930386 0.0082253052 0.006658891  0.0130892775
#> [25,]  0.008261643 0.0111780149 0.0100734547 0.007569317  0.0073011605
#>               V58         V59         V6           V60         V7          V8
#>  [1,] 0.017301726 0.023952093 0.05437482  0.0147235808 0.09369685  0.02860925
#>  [2,] 0.004390379 0.001300049 0.07568127  0.0006292036 0.16130300  0.15052415
#>  [3,] 0.008759632 0.012323110 0.14681005  0.0072782583 0.13244350  0.11068254
#>  [4,] 0.009661098 0.009287063 0.10740217  0.0060316167 0.09239933  0.12667462
#>  [5,] 0.004032326 0.000459539 0.02247022  0.0033958058 0.02880010 -0.03067679
#>  [6,] 0.010398441 0.006241978 0.04919221  0.0025830423 0.09818140  0.12506326
#>  [7,] 0.004345433 0.005040288 0.12132736  0.0038621538 0.13997624  0.16604225
#>  [8,] 0.015890384 0.012336067 0.16046181  0.0106254204 0.20768796  0.19385387
#>  [9,] 0.018552536 0.020293654 0.25992305  0.0100139066 0.23200593  0.15976468
#> [10,] 0.006712269 0.008158410 0.05077796  0.0036542135 0.04307121  0.17830352
#> [11,] 0.005579961 0.004363094 0.11000669  0.0026304935 0.09956887  0.10551520
#> [12,] 0.003347398 0.003417869 0.06970184  0.0047976826 0.10001179  0.13036757
#> [13,] 0.008686641 0.002375619 0.03856139  0.0097184000 0.05297237  0.10688970
#> [14,] 0.003700437 0.011356090 0.25799051  0.0077023111 0.23069350  0.21063803
#> [15,] 0.010097208 0.003425889 0.05050812  0.0034898479 0.21635736  0.31172183
#> [16,] 0.001566119 0.004063238 0.07393665  0.0050913242 0.11802734  0.07860186
#> [17,] 0.004279852 0.005444275 0.06502261  0.0050479926 0.05212668  0.05644623
#> [18,] 0.012034036 0.007236611 0.08087945  0.0095870646 0.11405962  0.10073939
#> [19,] 0.016651256 0.013072895 0.11175970  0.0141515924 0.22371922  0.23457994
#> [20,] 0.003672601 0.010077284 0.04171021  0.0051257517 0.05518351  0.05867642
#> [21,] 0.008009503 0.010917216 0.06653200  0.0069929445 0.13418314  0.18809995
#> [22,] 0.004170720 0.003213474 0.08690464  0.0038668887 0.08997876  0.07916093
#> [23,] 0.006428046 0.004811326 0.06579911 -0.0007628530 0.05119019  0.06057391
#> [24,] 0.006191818 0.006031805 0.13025985  0.0042173213 0.10874876  0.06556905
#> [25,] 0.010351453 0.010902880 0.08777015  0.0076550843 0.11241524  0.10787514
#>                V9
#>  [1,] 0.122305760
#>  [2,] 0.203061568
#>  [3,] 0.136644711
#>  [4,] 0.257125515
#>  [5,] 0.092232689
#>  [6,] 0.076155069
#>  [7,] 0.235907977
#>  [8,] 0.232476544
#>  [9,] 0.078335336
#> [10,] 0.205532290
#> [11,] 0.186188983
#> [12,] 0.201656737
#> [13,] 0.224296226
#> [14,] 0.346642715
#> [15,] 0.286964210
#> [16,] 0.095396104
#> [17,] 0.089321662
#> [18,] 0.064451641
#> [19,] 0.020312580
#> [20,] 0.089759268
#> [21,] 0.196167883
#> [22,] 0.151697899
#> [23,] 0.028019756
#> [24,] 0.007905115
#> [25,] 0.160435999
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3333333