Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11          V12        V13        V14
#>  [1,] 0.044127207 0.04012490 0.15036166  0.224394998 0.26532529 0.29258286
#>  [2,] 0.030700000 0.27920000 0.38770000  0.499200000 0.49810000 0.49720000
#>  [3,] 0.036412386 0.24640489 0.23438500  0.255541454 0.28896390 0.26309638
#>  [4,] 0.162018113 0.14334821 0.41243754  0.583002374 0.75771259 0.75614915
#>  [5,] 0.007369626 0.08635465 0.23745078  0.419346188 0.39296440 0.48776346
#>  [6,] 0.057337185 0.25788933 0.35380567  0.446263794 0.53699082 0.49818702
#>  [7,] 0.016995699 0.14750804 0.04209400 -0.019470415 0.10282845 0.25575035
#>  [8,] 0.031692386 0.26480966 0.31294723  0.341006541 0.41143792 0.53787036
#>  [9,] 0.105673377 0.56967171 0.47455556  0.214439056 0.20195940 0.28312525
#> [10,] 0.036646445 0.33377768 0.41905347  0.460314287 0.44352505 0.36622524
#> [11,] 0.029599173 0.24952389 0.29392282  0.317007912 0.26872160 0.17801107
#> [12,] 0.012146310 0.21225973 0.22547115  0.185687265 0.09011044 0.16468359
#> [13,] 0.026186708 0.25130110 0.23241779  0.163107157 0.16809218 0.25728699
#> [14,] 0.029300000 0.08600000 0.04140000  0.047200000 0.08350000 0.09380000
#> [15,] 0.018259521 0.14094622 0.14945439  0.150723572 0.12958895 0.15449778
#> [16,] 0.019863127 0.09353291 0.04410888  0.004053591 0.11652787 0.31264455
#> [17,] 0.015344191 0.15875170 0.15902848  0.133325082 0.15008710 0.14539703
#> [18,] 0.015034323 0.07001262 0.11919046  0.103381022 0.19584874 0.24729870
#> [19,] 0.007898993 0.07639157 0.05120806  0.131426515 0.12607361 0.13936451
#> [20,] 0.032134368 0.04411965 0.08213480  0.239572173 0.13429884 0.11281146
#> [21,] 0.007490284 0.15037710 0.11276225  0.102561201 0.23310378 0.31673124
#> [22,] 0.012831597 0.11526000 0.18406673  0.330216338 0.31628600 0.24857705
#> [23,] 0.028719508 0.23360737 0.22886630  0.155442389 0.21589264 0.34105309
#> [24,] 0.014563449 0.06807113 0.07158349  0.057812287 0.08419441 0.10355508
#> [25,] 0.011765560 0.10785899 0.09595711  0.010192779 0.03386179 0.05107191
#>             V15       V16        V17       V18       V19          V2       V20
#>  [1,] 0.3562304 0.4387991 0.60513159 0.7281849 0.8652162 0.044043870 0.9372830
#>  [2,] 0.5607000 0.7339000 0.82300000 0.9173000 0.9975000 0.052300000 0.9911000
#>  [3,] 0.1853078 0.2151283 0.24029456 0.1996375 0.2414552 0.052482212 0.2324823
#>  [4,] 0.7070683 0.8065929 0.96969014 0.9035952 0.9480551 0.142429917 0.7931185
#>  [5,] 0.6219608 0.6242003 0.76585477 0.9010540 0.9922398 0.015059557 1.0097909
#>  [6,] 0.5865915 0.7055401 0.89505195 0.9163885 0.8832505 0.111579937 0.7891557
#>  [7,] 0.4200481 0.5952708 0.70551698 0.7663014 0.8610905 0.006024952 0.9427103
#>  [8,] 0.6229024 0.6476407 0.65627282 0.7462816 0.8412974 0.044254147 0.8712985
#>  [9,] 0.5642264 0.5881712 0.17981162 0.1088852 0.3516899 0.105671118 0.8687390
#> [10,] 0.2793399 0.2971739 0.32557216 0.3510652 0.4365216 0.049453792 0.6388868
#> [11,] 0.1881535 0.2186712 0.22936288 0.3056346 0.3805946 0.039154947 0.4341125
#> [12,] 0.2702208 0.5161020 0.67247552 0.7800188 0.8046779 0.018465103 0.7082913
#> [13,] 0.4034245 0.5501815 0.66597078 0.5830043 0.6513059 0.035575281 0.8298093
#> [14,] 0.1466000 0.0809000 0.11790000 0.2179000 0.3326000 0.064400000 0.3258000
#> [15,] 0.1442699 0.1482396 0.19007153 0.2503321 0.2818031 0.016663965 0.5235351
#> [16,] 0.5297315 0.7547382 0.81675683 0.8062285 0.8483578 0.013931673 0.8634239
#> [17,] 0.2137171 0.2482133 0.26760545 0.2190688 0.2863900 0.014203487 0.3687200
#> [18,] 0.2502234 0.2194375 0.24393135 0.3037889 0.4667963 0.006508832 0.5394661
#> [19,] 0.1326694 0.1038184 0.07312003 0.1619363 0.1189694 0.025460590 0.1966461
#> [20,] 0.1074589 0.1610335 0.14039872 0.1021823 0.1697263 0.035719667 0.1321215
#> [21,] 0.3920694 0.3214605 0.12834776 0.1447488 0.1950639 0.024139034 0.2302754
#> [22,] 0.2282519 0.4831670 0.71466151 0.7110273 0.5300421 0.029281271 0.3826035
#> [23,] 0.4466295 0.6376302 0.70043876 0.5720255 0.4045885 0.027292682 0.3648414
#> [24,] 0.1359577 0.2211992 0.25813789 0.3163050 0.3903412 0.014049687 0.3747736
#> [25,] 0.1365791 0.0813183 0.07763639 0.1673143 0.2773769 0.013898880 0.3080903
#>             V21        V22       V23       V24        V25       V26       V27
#>  [1,] 0.9285625 0.93817119 0.9453660 0.9133370 0.80063709 0.6425911 0.5088578
#>  [2,] 0.8240000 0.64980000 0.5980000 0.4862000 0.31500000 0.1543000 0.0989000
#>  [3,] 0.2323523 0.28613383 0.4259632 0.5054213 0.58919459 0.6553995 0.6927931
#>  [4,] 0.8276279 0.82080888 0.7782096 0.6766492 0.45983087 0.4569692 0.3980776
#>  [5,] 0.9009399 0.84530450 0.8736937 0.8524010 0.59546257 0.4457401 0.4192630
#>  [6,] 0.7471839 0.50134952 0.2563212 0.2795004 0.23239599 0.1794237 0.1942250
#>  [7,] 0.9347112 0.93055954 0.8291520 0.5563088 0.35981426 0.3586630 0.2804795
#>  [8,] 0.8959730 0.93023839 0.9564093 0.9369357 0.83615420 0.8110059 0.7673455
#>  [9,] 0.8761702 0.40515061 0.1447246 0.1684903 0.23091551 0.3681958 0.2692285
#> [10,] 0.7398097 0.76855702 0.7576984 0.7419678 0.72590561 0.8078806 0.8484704
#> [11,] 0.5484579 0.69116260 0.7828429 0.8568653 0.90622937 0.9368092 0.9356119
#> [12,] 0.7180336 0.77787418 0.8112762 0.4873866 0.17226311 0.2873493 0.6184349
#> [13,] 0.9842388 1.00228975 0.7964442 0.5401952 0.37719759 0.4608352 0.4875620
#> [14,] 0.2111000 0.23020000 0.3361000 0.4259000 0.46090000 0.2606000 0.0874000
#> [15,] 0.6444091 0.79612301 0.9148229 0.8737120 0.88704515 0.9181111 0.8703683
#> [16,] 0.9452863 0.93093241 0.9237470 0.9570835 0.93601493 0.9336282 0.9228026
#> [17,] 0.4123305 0.39930394 0.5991084 0.7423331 0.79323602 0.8301881 0.7587294
#> [18,] 0.5533550 0.62353952 0.6543979 0.7102178 0.86319189 0.9370683 1.0094883
#> [19,] 0.2438428 0.20664604 0.1380416 0.1648336 0.15781450 0.3636748 0.4477758
#> [20,] 0.2003277 0.02406746 0.2262226 0.2152897 0.01327743 0.1873744 0.4706433
#> [21,] 0.2383636 0.32082139 0.4584361 0.6761205 0.68439583 0.7387349 0.7132026
#> [22,] 0.4084362 0.45490407 0.5727058 0.7338124 0.84981178 0.8818701 0.9030989
#> [23,] 0.4442479 0.56610832 0.6776708 0.7697697 0.80960975 0.7905561 0.7822094
#> [24,] 0.3595655 0.33710847 0.3328861 0.4420858 0.43628068 0.4703675 0.6334834
#> [25,] 0.3710378 0.71566306 0.8333012 0.8038503 0.70836534 0.7599412 0.6874627
#>             V28       V29          V3       V30       V31       V32        V33
#>  [1,] 0.3873426 0.3550256 0.075713339 0.3570697 0.3068534 0.2401155 0.32513913
#>  [2,] 0.0284000 0.1008000 0.065300000 0.2636000 0.2694000 0.2930000 0.29250000
#>  [3,] 0.8444575 0.8813490 0.051246538 0.8760307 0.9111607 0.8195508 0.65262498
#>  [4,] 0.5610503 0.4671864 0.148549685 0.2761462 0.1796356 0.1972132 0.28362073
#>  [5,] 0.4501431 0.4343113 0.043066274 0.3579124 0.2790874 0.3010372 0.30408683
#>  [6,] 0.2052300 0.2063352 0.125804847 0.1459356 0.2682687 0.3046521 0.33952476
#>  [7,] 0.4128795 0.5948393 0.005361828 0.5944440 0.4182651 0.3757668 0.44651848
#>  [8,] 0.6529469 0.5860281 0.049974581 0.4372338 0.3303300 0.2005168 0.09577189
#>  [9,] 0.5496518 0.5249344 0.049341629 0.6859913 0.3866584 0.4324589 0.64766697
#> [10,] 0.7633685 0.6029285 0.053399991 0.5159800 0.4637213 0.4713200 0.44349254
#> [11,] 0.9080275 0.8238520 0.043354302 0.6735741 0.5994117 0.4751771 0.34969234
#> [12,] 0.8529748 0.8961426 0.004548658 0.7802218 0.5244684 0.5182647 0.42646474
#> [13,] 0.2730244 0.1736538 0.018182082 0.4160660 0.2648383 0.4331576 0.45450949
#> [14,] 0.2862000 0.5606000 0.039000000 0.8344000 0.8096000 0.7250000 0.80480000
#> [15,] 0.7884150 0.7287737 0.021707671 0.6651163 0.6679385 0.5528387 0.52667430
#> [16,] 0.7192584 0.4204500 0.027573401 0.3299963 0.3803610 0.3976856 0.33246300
#> [17,] 0.7249436 0.5518826 0.024327872 0.3373114 0.2658296 0.1684002 0.13108439
#> [18,] 0.9212995 0.7547772 0.010027415 0.7148636 0.6093850 0.4070438 0.30803478
#> [19,] 0.4599730 0.4896265 0.020683662 0.5692778 0.6933162 0.7336225 0.73092181
#> [20,] 0.6950474 0.6829771 0.034679079 0.6482974 0.7172160 0.5401534 0.64939247
#> [21,] 0.8394276 0.9274485 0.040015025 0.8725151 0.7889033 0.6535373 0.55671099
#> [22,] 0.8942492 0.9175931 0.027929854 0.8617827 0.7605154 0.6902840 0.73154980
#> [23,] 0.7333135 0.5629717 0.043746144 0.5636376 0.6153039 0.5086455 0.41572674
#> [24,] 0.8706082 1.0017486 0.017432043 0.9321778 0.8438280 0.7521543 0.71732565
#> [25,] 0.7445324 0.8417959 0.033118766 0.7985663 0.6121588 0.5959685 0.76889973
#>             V34       V35        V36       V37       V38       V39         V4
#>  [1,] 0.4078117 0.3996652 0.37983555 0.4086209 0.4372156 0.4154603 0.08396390
#>  [2,] 0.3998000 0.3660000 0.31720000 0.4609000 0.4374000 0.1820000 0.05210000
#>  [3,] 0.5396688 0.5184982 0.45889298 0.4148174 0.4189036 0.4265614 0.05363056
#>  [4,] 0.2806489 0.3191816 0.06558414 0.1765082 0.3161194 0.3347364 0.16074981
#>  [5,] 0.2027858 0.2072244 0.24748920 0.3438117 0.4998700 0.3990169 0.12031449
#>  [6,] 0.4413821 0.5103286 0.25319881 0.1556797 0.2418867 0.2952235 0.15451695
#>  [7,] 0.3533387 0.2704717 0.18667582 0.1000402 0.1169091 0.1124624 0.05056789
#>  [8,] 0.1071542 0.1641678 0.17001911 0.1854362 0.1992831 0.2696827 0.05608736
#>  [9,] 0.7545714 0.8883526 0.85966130 0.8689310 0.8906450 0.7277000 0.08702480
#> [10,] 0.3834415 0.2369873 0.31568799 0.3346596 0.3581144 0.4607632 0.05093159
#> [11,] 0.2105647 0.1281911 0.11269301 0.1718711 0.1661265 0.1697131 0.05202808
#> [12,] 0.3501211 0.2830024 0.23569899 0.2831446 0.4674194 0.4502936 0.02016242
#> [13,] 0.2252032 0.1232970 0.33050542 0.4262127 0.3927671 0.2877768 0.02975961
#> [14,] 0.9435000 1.0000000 0.89600000 0.5516000 0.3037000 0.2338000 0.01730000
#> [15,] 0.5307699 0.5234651 0.46336540 0.4527219 0.3004516 0.1094751 0.03654693
#> [16,] 0.2836079 0.2436212 0.13879432 0.2131387 0.1790489 0.1117407 0.02677631
#> [17,] 0.2477012 0.3030246 0.24268561 0.1565559 0.1457385 0.1691344 0.02380323
#> [18,] 0.1637200 0.1637233 0.33481894 0.2944348 0.2993634 0.3072639 0.01803331
#> [19,] 0.7765436 0.8950250 0.86481315 0.8393159 0.7819937 0.7304967 0.02308626
#> [20,] 0.8763272 0.9630041 0.97044228 0.7754534 0.6694267 0.6324227 0.02541439
#> [21,] 0.5138892 0.4789606 0.43641434 0.2241126 0.2017705 0.2358698 0.02478243
#> [22,] 0.8497879 0.9632346 0.98454856 0.9286371 0.8011895 0.6316635 0.03461544
#> [23,] 0.4068545 0.6231286 0.67786855 0.5919596 0.3895966 0.2762475 0.04552386
#> [24,] 0.6270469 0.5475793 0.55872173 0.5492567 0.3590871 0.1713604 0.01666488
#> [25,] 0.9060034 0.8723849 0.68402734 0.4154414 0.2381885 0.1841204 0.03135620
#>              V40        V41        V42        V43        V44         V45
#>  [1,] 0.35190589 0.33542082 0.39634043 0.31006369 0.20241262 0.148079140
#>  [2,] 0.33760000 0.62020000 0.44480000 0.18630000 0.14200000 0.058900000
#>  [3,] 0.42256369 0.51503360 0.52703378 0.49879097 0.51993779 0.577422350
#>  [4,] 0.21427410 0.06006687 0.28098101 0.41035223 0.16168089 0.079093196
#>  [5,] 0.26392996 0.15297043 0.16801818 0.24136182 0.10904799 0.040855257
#>  [6,] 0.17940068 0.07646166 0.03617868 0.08253516 0.06990184 0.040270562
#>  [7,] 0.07074561 0.17605028 0.22176264 0.13562857 0.19466427 0.216560464
#>  [8,] 0.24447811 0.22005126 0.21620799 0.22957248 0.18717492 0.117339073
#>  [9,] 0.49225045 0.26095006 0.24461069 0.23970776 0.30772975 0.461211557
#> [10,] 0.40391760 0.31114104 0.37582637 0.40231045 0.35257627 0.365265083
#> [11,] 0.14619613 0.13640355 0.17242507 0.19242342 0.17618270 0.154591219
#> [12,] 0.33649932 0.11546141 0.18259696 0.40281036 0.32592347 0.090743629
#> [13,] 0.19719980 0.11580464 0.21713792 0.17118469 0.10586284 0.060065193
#> [14,] 0.23820000 0.33180000 0.38210000 0.15750000 0.22280000 0.158200000
#> [15,] 0.14141175 0.15800207 0.16627143 0.16680620 0.05523867 0.001992491
#> [16,] 0.11656384 0.06600371 0.02751002 0.05056228 0.10419320 0.092955448
#> [17,] 0.13607536 0.15019177 0.13790095 0.12968499 0.10957547 0.097264925
#> [18,] 0.15978546 0.10721887 0.23628657 0.21585446 0.14545048 0.139454296
#> [19,] 0.62568812 0.43600330 0.37060403 0.27676151 0.22782068 0.236238211
#> [20,] 0.66366229 0.58891131 0.50938116 0.37179132 0.21788636 0.141968490
#> [21,] 0.17937064 0.14108257 0.10755503 0.18215072 0.14840121 0.035602814
#> [22,] 0.45774223 0.29377890 0.19424119 0.15459126 0.11240415 0.177223980
#> [23,] 0.32872207 0.30103890 0.27214338 0.22597464 0.20854986 0.180804524
#> [24,] 0.11002893 0.07313388 0.14357737 0.22570635 0.28372808 0.248669298
#> [25,] 0.28924667 0.28999488 0.25881566 0.33876447 0.25070997 0.133614127
#>               V46          V47        V48         V49            V5         V50
#>  [1,] 0.134143248  0.093686785 0.04940478 0.037754647  1.410502e-01 0.014571691
#>  [2,] 0.057600000  0.067200000 0.02690000 0.024500000  6.110000e-02 0.019000000
#>  [3,] 0.488421427  0.334126000 0.22693495 0.119494455  7.195392e-02 0.035429809
#>  [4,] 0.118541619  0.059286537 0.05715750 0.060685608  1.964993e-01 0.038866370
#>  [5,] 0.004589993 -0.007017407 0.03631502 0.005675705  1.982802e-01 0.004313935
#>  [6,] 0.050751556  0.102522651 0.11080352 0.074749403  1.360284e-01 0.026670315
#>  [7,] 0.101619366  0.071855101 0.06466171 0.043301868  7.273062e-02 0.010169114
#>  [8,] 0.104641826  0.091724260 0.06870537 0.048842661  9.094437e-02 0.021265781
#>  [9,] 0.335687575  0.137881668 0.07957285 0.049282062  8.335294e-02 0.026586642
#> [10,] 0.230011781  0.146518090 0.11885629 0.077941085  7.460364e-02 0.027163520
#> [11,] 0.146727607  0.151226893 0.11899942 0.063956893  6.614792e-02 0.019352104
#> [12,] 0.166849313  0.145540596 0.07857215 0.064477849  2.081041e-02 0.022902685
#> [13,] 0.025084714  0.050683728 0.06625610 0.039843734  4.457778e-02 0.020368940
#> [14,] 0.143300000  0.163400000 0.11330000 0.056700000  4.760000e-02 0.013300000
#> [15,] 0.031988415  0.037220984 0.04108097 0.015818970  7.888708e-02 0.006641731
#> [16,] 0.040953457  0.045492046 0.03793477 0.022002378  6.608659e-02 0.006078109
#> [17,] 0.092822492  0.042185351 0.02231795 0.018708177  4.906841e-02 0.012468522
#> [18,] 0.069826592  0.075013506 0.05066678 0.031063439  4.088272e-02 0.021568759
#> [19,] 0.160385111  0.120573354 0.09182850 0.040571147 -7.908712e-05 0.018983826
#> [20,] 0.070362028  0.078998714 0.07656237 0.055789985  6.618246e-02 0.017523129
#> [21,] 0.031820821  0.027863926 0.01673833 0.001934277  4.504600e-02 0.017463488
#> [22,] 0.192086653  0.110858557 0.06005708 0.018422432  3.055766e-02 0.024289967
#> [23,] 0.159310874  0.114109305 0.09198883 0.057315985  6.714773e-02 0.019849610
#> [24,] 0.168396316  0.090797623 0.08839295 0.053378391  1.477762e-02 0.015482421
#> [25,] 0.074721047  0.053177021 0.03405173 0.009489855  5.541590e-02 0.013203396
#>                V51         V52         V53         V54         V55         V56
#>  [1,] 0.0117760167 0.010719039 0.011745238 0.019350571 0.010212717 0.010201117
#>  [2,] 0.0063000000 0.032100000 0.018900000 0.013700000 0.027700000 0.015200000
#>  [3,] 0.0277758846 0.026518286 0.015947138 0.010581816 0.008190737 0.008540750
#>  [4,] 0.0209674929 0.017830766 0.018999328 0.040787860 0.008977825 0.007224712
#>  [5,] 0.0022563634 0.019678042 0.011019179 0.009794973 0.030095590 0.030833647
#>  [6,] 0.0234356515 0.033870825 0.013400940 0.009208621 0.016802063 0.013387086
#>  [7,] 0.0151984534 0.011286438 0.002325330 0.002561878 0.001931748 0.002404888
#>  [8,] 0.0252943588 0.017073111 0.013475999 0.021580232 0.010117968 0.009597034
#>  [9,] 0.0173019149 0.021531862 0.012586474 0.021682480 0.022193287 0.010468262
#> [10,] 0.0195923005 0.013679856 0.009281070 0.009241018 0.010556496 0.006663237
#> [11,] 0.0158962080 0.013323421 0.007694906 0.009621855 0.005116738 0.006381125
#> [12,] 0.0110610739 0.017812079 0.009222818 0.003750672 0.005333894 0.006587584
#> [13,] 0.0141851866 0.010851083 0.001544086 0.010698898 0.003186239 0.010658420
#> [14,] 0.0170000000 0.003500000 0.005200000 0.008300000 0.007800000 0.007500000
#> [15,] 0.0025201594 0.001688745 0.010013697 0.010362076 0.011820622 0.005894132
#> [16,] 0.0146586945 0.011448381 0.006441289 0.005663509 0.004696021 0.007877710
#> [17,] 0.0125684279 0.017276174 0.007197746 0.004850773 0.007442824 0.005837929
#> [18,] 0.0005204466 0.006417525 0.005256178 0.006568500 0.006959499 0.005161548
#> [19,] 0.0204022613 0.008570462 0.008464220 0.006682872 0.014861300 0.009660469
#> [20,] 0.0056584250 0.008066495 0.016445745 0.016236740 0.008124630 0.006376555
#> [21,] 0.0145694295 0.014725489 0.015397074 0.005187110 0.004780225 0.005951608
#> [22,] 0.0242275810 0.017654365 0.008135862 0.007978901 0.009510168 0.007417602
#> [23,] 0.0121475264 0.010261765 0.006004018 0.009355986 0.008698363 0.005706743
#> [24,] 0.0094001182 0.006301060 0.004042874 0.010384408 0.005291565 0.003757159
#> [25,] 0.0172905602 0.014888263 0.017479643 0.013866228 0.005070147 0.010930429
#>                V57         V58          V59         V6           V60         V7
#>  [1,]  0.011758430 0.011300670 0.0121591116 0.10209930  7.447253e-03 0.07512137
#>  [2,]  0.005200000 0.012100000 0.0124000000 0.05770000  5.500000e-03 0.06650000
#>  [3,]  0.008591064 0.009299529 0.0078161593 0.11300914  4.197413e-03 0.10490317
#>  [4,] -0.001081411 0.009455307 0.0175576304 0.15615969  5.897494e-03 0.17750565
#>  [5,]  0.010348355 0.005770102 0.0164567005 0.08961372  8.976826e-03 0.06900239
#>  [6,]  0.003589165 0.014344532 0.0226244734 0.06156338  1.325630e-02 0.09580722
#>  [7,]  0.001984916 0.004518492 0.0036473935 0.14934018 -6.155699e-05 0.14732932
#>  [8,]  0.009780355 0.013319646 0.0084898021 0.16894092  7.130343e-03 0.18490827
#>  [9,]  0.009819628 0.015914929 0.0133685419 0.12784672  1.284485e-02 0.24522976
#> [10,]  0.005965674 0.007148298 0.0075097231 0.12196049  5.104257e-03 0.11501345
#> [11,]  0.005477497 0.005475432 0.0066353641 0.11092453  7.014622e-03 0.12731602
#> [12,]  0.003038926 0.006960067 0.0021701352 0.03865068  4.229532e-03 0.05990438
#> [13,]  0.005818414 0.008512025 0.0003045274 0.02376295  8.035187e-03 0.05211755
#> [14,]  0.010500000 0.016000000 0.0095000000 0.08160000  1.100000e-03 0.09930000
#> [15,]  0.005975093 0.004238658 0.0024862626 0.12290960  6.386115e-03 0.11477848
#> [16,]  0.006977152 0.003545170 0.0067751828 0.11694091  4.772115e-03 0.11496504
#> [17,]  0.003377990 0.005672075 0.0035556968 0.07653521  3.670375e-03 0.07742850
#> [18,]  0.005330057 0.005143976 0.0012717222 0.01786479  3.822209e-03 0.06969903
#> [19,]  0.002298138 0.003263853 0.0024558921 0.05102260  3.115655e-04 0.06142790
#> [20,]  0.012772974 0.008871853 0.0115585830 0.11038007  5.281595e-03 0.09461751
#> [21,]  0.005256508 0.002938952 0.0115892586 0.10995035  7.998527e-03 0.09977075
#> [22,]  0.008119879 0.009763444 0.0067180045 0.10341752  1.015376e-02 0.14385546
#> [23,]  0.013269434 0.008648968 0.0107696211 0.11849146  8.050221e-03 0.14414276
#> [24,]  0.004914839 0.004059331 0.0076218634 0.08835365  4.420923e-03 0.09446953
#> [25,]  0.016824323 0.006038570 0.0070748435 0.05243843  1.133347e-02 0.08875642
#>               V8          V9
#>  [1,] 0.08318079 0.097231231
#>  [2,] 0.06640000 0.146000000
#>  [3,] 0.14551367 0.248145573
#>  [4,] 0.18157906 0.094958483
#>  [5,] 0.10846147 0.063442746
#>  [6,] 0.07195519 0.154751612
#>  [7,] 0.14361392 0.199132890
#>  [8,] 0.20747158 0.255988897
#>  [9,] 0.44301484 0.586568059
#> [10,] 0.16350429 0.263626775
#> [11,] 0.13950822 0.181285452
#> [12,] 0.02497954 0.142923818
#> [13,] 0.09527063 0.247430975
#> [14,] 0.03150000 0.073600000
#> [15,] 0.12057370 0.122029215
#> [16,] 0.06647037 0.106218775
#> [17,] 0.07783995 0.112754993
#> [18,] 0.06297172 0.100035928
#> [19,] 0.08307259 0.006238036
#> [20,] 0.06310222 0.013608654
#> [21,] 0.12833385 0.163156108
#> [22,] 0.08660884 0.075375209
#> [23,] 0.17220221 0.181068775
#> [24,] 0.05306393 0.072443735
#> [25,] 0.06658653 0.020553950
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623