Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.096524471 0.23986019 0.35313233 0.43955085 0.57354301 0.7184639
#>  [2,] 0.053542830 0.39033403 0.53049318 0.51516739 0.55738371 0.5892623
#>  [3,] 0.032990499 0.31559242 0.39545648 0.47002663 0.45585394 0.3947857
#>  [4,] 0.026351352 0.26954483 0.31291389 0.31053760 0.26141737 0.1410536
#>  [5,] 0.059272680 0.23297766 0.31349192 0.41589178 0.51275773 0.4790788
#>  [6,] 0.020231857 0.23326243 0.27101117 0.30221000 0.27918542 0.2110397
#>  [7,] 0.028772961 0.19021415 0.25162500 0.29502913 0.24150587 0.2500851
#>  [8,] 0.014369055 0.23227813 0.25712684 0.32346512 0.32774681 0.2918106
#>  [9,] 0.045037346 0.26441948 0.25703960 0.25193278 0.25127151 0.2562049
#> [10,] 0.016200000 0.13340000 0.29690000 0.47540000 0.56770000 0.5690000
#> [11,] 0.025829219 0.27566832 0.31877943 0.31064068 0.30315451 0.3085017
#> [12,] 0.021702632 0.15082143 0.23268082 0.25691744 0.27247248 0.2832811
#> [13,] 0.044114822 0.02714833 0.14114858 0.22986756 0.28276197 0.3060551
#> [14,] 0.018729045 0.14512326 0.14215708 0.11028049 0.08378494 0.1528172
#> [15,] 0.028534210 0.22772927 0.23689867 0.13194093 0.18366122 0.4837168
#> [16,] 0.012786199 0.08841829 0.09653879 0.25442439 0.09732356 0.2137166
#> [17,] 0.005679762 0.23617512 0.14874204 0.07325950 0.23785514 0.3314494
#> [18,] 0.031160327 0.33838378 0.16648104 0.05979934 0.05425309 0.3805425
#> [19,] 0.024517319 0.06525329 0.13392016 0.04444068 0.08639231 0.1319757
#> [20,] 0.017365194 0.12361942 0.14120756 0.11599920 0.13266352 0.0964277
#> [21,] 0.013350341 0.12255106 0.12344185 0.15576072 0.15677045 0.1280496
#> [22,] 0.018654408 0.06814717 0.09446717 0.18247585 0.16386460 0.1610754
#> [23,] 0.027711439 0.35123427 0.29121609 0.05872548 0.26053162 0.5249010
#> [24,] 0.008071896 0.08886523 0.05269699 0.06417870 0.11474145 0.1501310
#> [25,] 0.014453460 0.10715337 0.14319891 0.14983025 0.18527253 0.1868333
#>              V15        V16        V17       V18        V19           V2
#>  [1,] 0.93048102 0.91491489 0.77656655 0.8763055 1.00200665  0.114209876
#>  [2,] 0.69624315 0.81847357 0.94013212 0.9392245 0.83235772  0.076044455
#>  [3,] 0.35680719 0.39290421 0.46933120 0.5356333 0.61121202  0.044810412
#>  [4,] 0.15579394 0.20254703 0.23541813 0.2631425 0.23197688  0.025481605
#>  [5,] 0.58852957 0.71259949 0.89979294 0.9121129 0.87528734  0.123936055
#>  [6,] 0.17464407 0.19265233 0.22831340 0.3380830 0.52070664  0.030900032
#>  [7,] 0.22241556 0.21922913 0.18368190 0.2764111 0.44042535  0.042959836
#>  [8,] 0.22444813 0.23530770 0.21673089 0.2924886 0.49294643  0.017140072
#>  [9,] 0.27230159 0.31965828 0.28763615 0.2182792 0.26791800  0.061508307
#> [10,] 0.64210000 0.74870000 0.89990000 1.0000000 0.96900000  0.025300000
#> [11,] 0.25968843 0.24257824 0.25821081 0.2850104 0.42818360  0.037178285
#> [12,] 0.23299975 0.18013319 0.16185780 0.2193095 0.34739011  0.043955997
#> [13,] 0.44863764 0.49390597 0.65097101 0.7111286 0.81990381  0.041823826
#> [14,] 0.08570476 0.10833241 0.24767313 0.2832514 0.32239922  0.008666385
#> [15,] 0.70483882 0.74742503 0.77202845 0.6144566 0.42021498  0.036651781
#> [16,] 0.16737048 0.23328657 0.17381390 0.3311605 0.44404113  0.015725258
#> [17,] 0.44975361 0.27447854 0.05888279 0.1389216 0.24227869  0.021692100
#> [18,] 0.54275993 0.54810800 0.52337885 0.4360926 0.21317755  0.094805153
#> [19,] 0.16977499 0.18405849 0.16029992 0.1342811 0.08453115  0.017471364
#> [20,] 0.10689456 0.09915411 0.15369369 0.2828359 0.35469134  0.022483502
#> [21,] 0.11760804 0.28587641 0.42465455 0.4313031 0.39492310  0.019607904
#> [22,] 0.15898806 0.20750627 0.27892627 0.3321103 0.35669367  0.032749359
#> [23,] 0.62218471 0.78232064 0.89018632 0.7848353 0.53234524  0.114076578
#> [24,] 0.22937553 0.28958412 0.36145091 0.3637294 0.30194765 -0.006196806
#> [25,] 0.19541090 0.23974141 0.30632793 0.3324864 0.39044435  0.012247436
#>             V20       V21       V22        V23        V24       V25       V26
#>  [1,] 0.9366509 0.8880768 0.8975750 1.02243318 0.95057985 0.6544121 0.5581404
#>  [2,] 0.7557654 0.6911126 0.4202120 0.09799449 0.02146303 0.2447154 0.3330290
#>  [3,] 0.7237933 0.7651220 0.8663258 0.92582253 0.96738628 0.9470790 0.9284349
#>  [4,] 0.2359283 0.3952383 0.6424449 0.79576859 0.84152061 0.9058645 0.9404623
#>  [5,] 0.7680359 0.7416347 0.5272003 0.30602459 0.31402889 0.2228117 0.1573330
#>  [6,] 0.6405576 0.7526269 0.7986425 0.83044719 0.93098747 0.9538277 0.9809815
#>  [7,] 0.6003524 0.7874012 0.8408359 0.76613594 0.77133280 0.7495448 0.8405578
#>  [8,] 0.6212334 0.7591325 0.7780315 0.74332504 0.85966097 0.9192078 0.9771138
#>  [9,] 0.3840705 0.4445131 0.4533601 0.44993198 0.53576122 0.5736576 0.6577566
#> [10,] 0.9032000 0.7685000 0.6998000 0.66440000 0.59640000 0.3711000 0.0921000
#> [11,] 0.5524098 0.7041993 0.7194444 0.69072790 0.67923504 0.6449367 0.7118715
#> [12,] 0.3379909 0.3587863 0.3857419 0.47239999 0.61767435 0.7063498 0.7754315
#> [13,] 0.9224734 0.8971214 0.9217500 0.96455853 0.94720088 0.7756035 0.5310960
#> [14,] 0.4958144 0.6215347 0.7525148 0.88720054 0.90030789 0.9171383 0.9751754
#> [15,] 0.2824660 0.4958714 0.7442478 0.85468708 0.99076455 0.8538232 0.6782903
#> [16,] 0.3528008 0.2068197 0.3017395 0.50924126 0.42993098 0.2973508 0.4700396
#> [17,] 0.2726426 0.2617174 0.4005093 0.64753561 0.81299123 0.6960932 0.6707840
#> [18,] 0.4252465 0.7686186 0.9672806 0.93435854 0.55300200 0.5254271 0.6804374
#> [19,] 0.1616217 0.2950372 0.5381574 0.79337499 0.82918010 0.8122887 0.8494486
#> [20,] 0.4240405 0.5379171 0.7839165 0.87052287 0.84591733 0.7591560 0.7394197
#> [21,] 0.4311937 0.5375176 0.6626592 0.76147477 0.78734118 0.8095228 0.8911421
#> [22,] 0.3796690 0.3648973 0.2813878 0.35437613 0.45532014 0.3682723 0.3582859
#> [23,] 0.4153175 0.5489143 0.7116398 0.80850891 0.94894312 0.8460330 0.7189659
#> [24,] 0.2815307 0.4018578 0.4360343 0.46074454 0.60394173 0.7378973 0.9012302
#> [25,] 0.4528593 0.5008275 0.5704800 0.70324571 0.78863401 0.8318865 0.8265116
#>             V27       V28        V29         V3       V30       V31       V32
#>  [1,] 0.4167355 0.3540121 0.45192044 0.10657435 0.2097112 0.1875202 0.1081546
#>  [2,] 0.3127723 0.2717707 0.06800572 0.11652442 0.2241075 0.1971971 0.2122747
#>  [3,] 0.9003498 0.8745145 0.80618622 0.04180724 0.5769328 0.3605176 0.3064971
#>  [4,] 0.9751306 0.9966068 0.85341597 0.02532423 0.6777637 0.6521046 0.5612216
#>  [5,] 0.1680826 0.1795169 0.19464988 0.12861634 0.1658871 0.2723365 0.3110710
#>  [6,] 0.9782683 0.9184058 0.82393463 0.03348697 0.6648260 0.4769033 0.3441420
#>  [7,] 0.8882439 0.8474685 0.72037292 0.05056705 0.5266885 0.3655477 0.2763340
#>  [8,] 1.0000000 0.9611734 0.87216702 0.03174353 0.6792211 0.4875485 0.3371623
#>  [9,] 0.7278587 0.8173622 0.83076119 0.07167689 0.8865266 0.8269824 0.7488290
#> [10,] 0.0481000 0.0876000 0.10400000 0.02620000 0.1714000 0.3264000 0.4612000
#> [11,] 0.8042195 0.7915156 0.91622964 0.03753824 0.8661166 0.6827182 0.5480691
#> [12,] 0.8605379 0.9313947 0.87571542 0.04880406 0.6625998 0.5561492 0.4658470
#> [13,] 0.3868943 0.3307308 0.38860016 0.07886998 0.3951060 0.2310486 0.1832601
#> [14,] 0.9665308 0.8752861 0.80925184 0.01278062 0.7004173 0.6275065 0.5742571
#> [15,] 0.6985143 0.6023347 0.48388585 0.08394575 0.5359937 0.4432210 0.1741435
#> [16,] 0.5006524 0.3052998 0.31406855 0.06309195 0.4185238 0.6932174 0.7725475
#> [17,] 0.6771667 0.8453923 1.00287849 0.05135289 0.9588691 0.8937713 0.7489248
#> [18,] 0.5706084 0.5390698 0.20996409 0.13061144 0.2070268 0.5735425 0.6375305
#> [19,] 0.8128666 0.8550732 0.91355199 0.02342126 0.7567077 0.5971057 0.4762084
#> [20,] 0.6256233 0.6388739 0.70113028 0.03520531 0.6788986 0.5994787 0.6601078
#> [21,] 0.9590082 0.9329280 0.89309770 0.01874317 0.8182982 0.7616670 0.7216135
#> [22,] 0.4878661 0.6752407 0.73200716 0.03611155 0.6815390 0.7080297 0.6328823
#> [23,] 0.7053884 0.5045094 0.35255343 0.12146894 0.4112893 0.4229742 0.2083958
#> [24,] 0.9247120 0.6105985 0.26564789 0.01959063 0.3185481 0.5280285 0.5242397
#> [25,] 0.8739674 0.8538842 0.68717099 0.01801767 0.5529544 0.4625450 0.3801431
#>               V33        V34        V35        V36        V37        V38
#>  [1,] -0.01428364 0.07948291 0.21063837 0.14869311 0.15640045 0.07948842
#>  [2,]  0.27727676 0.21123845 0.30798609 0.05359829 0.10357107 0.22518784
#>  [3,]  0.28061152 0.17651749 0.10393444 0.14044447 0.18592784 0.28106663
#>  [4,]  0.44428715 0.20121693 0.09703089 0.07028898 0.11489398 0.15537382
#>  [5,]  0.34594733 0.45307484 0.52058376 0.27555353 0.15889438 0.27003540
#>  [6,]  0.22956348 0.20908817 0.20740130 0.16059280 0.16625471 0.17220686
#>  [7,]  0.25193516 0.22712068 0.15175034 0.06411343 0.08881530 0.14806641
#>  [8,]  0.31720025 0.20146437 0.10160988 0.04043766 0.07872247 0.10642914
#>  [9,]  0.57523769 0.50370882 0.52669528 0.45860640 0.47249571 0.50671708
#> [10,]  0.39390000 0.50500000 0.48330000 0.35110000 0.23190000 0.40290000
#> [11,]  0.56213395 0.48689033 0.28833831 0.30901833 0.25019756 0.19183155
#> [12,]  0.41987472 0.36269673 0.21225149 0.13112355 0.16662446 0.15676396
#> [13,]  0.30136302 0.40492023 0.43393720 0.33788668 0.29846650 0.36184852
#> [14,]  0.54150557 0.57145321 0.60125234 0.54018957 0.48035800 0.37176738
#> [15,]  0.18419016 0.08710681 0.42623117 0.70897361 0.72135958 0.33817310
#> [16,]  0.63668356 0.42316747 0.45380573 0.39433034 0.60395518 0.89438511
#> [17,]  0.68362220 0.55981551 0.42240916 0.43228967 0.26168163 0.12597872
#> [18,]  0.30161363 0.18320746 0.28950782 0.51938810 0.62617250 0.44016397
#> [19,]  0.53211175 0.69956489 0.69213706 0.51519408 0.23793067 0.13484223
#> [20,]  0.80554101 0.84342361 0.78662835 0.63060906 0.41901453 0.25651791
#> [21,]  0.70059155 0.72072458 0.75052052 0.70899231 0.64916693 0.55524348
#> [22,]  0.64621322 0.71585927 0.69823383 0.71419401 0.65148160 0.50039439
#> [23,]  0.18130801 0.11863795 0.45605018 0.67805287 0.63160567 0.37869080
#> [24,]  0.33048032 0.45975529 0.53961947 0.52366995 0.48019050 0.49425409
#> [25,]  0.27066318 0.16945702 0.15249799 0.23724080 0.23622588 0.24447333
#>              V39         V4        V40        V41        V42        V43
#>  [1,] 0.29175760 0.04469427 0.25962077 0.12253171 0.19702242 0.37873562
#>  [2,] 0.18687048 0.12810447 0.04947438 0.09648348 0.19260314 0.15726442
#>  [3,] 0.29043316 0.06859217 0.25471844 0.31921685 0.35370633 0.27918304
#>  [4,] 0.17165632 0.03570874 0.09781038 0.13156642 0.11287573 0.17139857
#>  [5,] 0.32560373 0.15325745 0.19314187 0.09023744 0.04370168 0.07534601
#>  [6,] 0.14581539 0.02875318 0.13044255 0.07868495 0.12570982 0.20513974
#>  [7,] 0.14718011 0.06096740 0.14185436 0.11362024 0.15598312 0.19479089
#>  [8,] 0.13001369 0.03706200 0.15432305 0.07279044 0.15964903 0.25129761
#>  [9,] 0.46588065 0.07602343 0.37088707 0.50495941 0.52180142 0.42970102
#> [10,] 0.36760000 0.03860000 0.15100000 0.07450000 0.13950000 0.15520000
#> [11,] 0.30522604 0.04695547 0.32512409 0.30343474 0.38537548 0.38224525
#> [12,] 0.22818503 0.07365297 0.31838055 0.31042731 0.32514190 0.22405845
#> [13,] 0.41206271 0.06731920 0.38094083 0.34203987 0.38806418 0.34799097
#> [14,] 0.18892097 0.01752825 0.16505999 0.18244657 0.10038972 0.10386694
#> [15,] 0.09509389 0.06851206 0.32164718 0.43716678 0.39492429 0.25144372
#> [16,] 0.95792725 0.01884687 0.89419228 0.65833664 0.54913697 0.36162528
#> [17,] 0.14931914 0.04182058 0.05292918 0.08681836 0.15259969 0.25812424
#> [18,] 0.54473069 0.14031347 0.60897576 0.50843468 0.24427682 0.19340190
#> [19,] 0.15389246 0.04203662 0.22183116 0.18589931 0.20959842 0.30027002
#> [20,] 0.25815197 0.03296544 0.39673390 0.40782572 0.35249152 0.37131679
#> [21,] 0.39706765 0.03269241 0.27796309 0.20069870 0.12690998 0.09791623
#> [22,] 0.40721334 0.04682892 0.37200210 0.33347023 0.29191231 0.18623779
#> [23,] 0.14151897 0.10108992 0.30065622 0.33528665 0.32261809 0.23342251
#> [24,] 0.45127755 0.01352606 0.44717119 0.44702731 0.29850286 0.22738450
#> [25,] 0.23410733 0.01630690 0.16796302 0.13116070 0.17254990 0.16316749
#>              V44         V45          V46         V47         V48          V49
#>  [1,] 0.26224312  0.07525709  0.116846445 0.080313748  0.03805217  0.049308130
#>  [2,] 0.11214529  0.09690664  0.046684394 0.066829712  0.06850339  0.021086893
#>  [3,] 0.24370147  0.24580692  0.175338602 0.095176042  0.07178571  0.046356367
#>  [4,] 0.14737806  0.08563387  0.074306131 0.114623257  0.12009880  0.073208782
#>  [5,] 0.06832800  0.03957687  0.056795807 0.110454827  0.11559667  0.078229997
#>  [6,] 0.20516507  0.17175210  0.187444529 0.141831161  0.07178060  0.042983019
#>  [7,] 0.17608990  0.18360892  0.216539900 0.178598399  0.10236878  0.057336723
#>  [8,] 0.28258875  0.27725576  0.225465829 0.177274275  0.11880508  0.055955318
#>  [9,] 0.44925479  0.51555357  0.431084808 0.265238658  0.17711570  0.098956304
#> [10,] 0.03770000  0.06360000  0.044300000 0.026400000  0.02230000  0.018700000
#> [11,] 0.20784324  0.19241537  0.104832924 0.104358952  0.12788572  0.077081281
#> [12,] 0.11595645  0.16473068  0.166238673 0.169041147  0.14470560  0.065068773
#> [13,] 0.27227125  0.19999590  0.142184639 0.099218629  0.04572759  0.032427121
#> [14,] 0.06120695  0.04070846  0.041243691 0.030939401  0.03786202  0.014191162
#> [15,] 0.19185996  0.22495690  0.201812648 0.197343821  0.17103164  0.127977297
#> [16,] 0.32164026  0.33771140  0.287344671 0.235195670  0.14662045  0.064572641
#> [17,] 0.14283548 -0.01745108 -0.004747007 0.018209260  0.03010326  0.001908421
#> [18,] 0.17390554  0.12993408  0.062364124 0.111462674  0.10220215  0.049650658
#> [19,] 0.16875231  0.03753114  0.038707532 0.047856856  0.04082186  0.008929721
#> [20,] 0.27306067  0.17696010  0.115666200 0.074124449  0.05365406  0.024735917
#> [21,] 0.08784346  0.06838936  0.084304493 0.049802925  0.04226607  0.015798264
#> [22,] 0.14899002  0.09358570  0.063381527 0.062880778  0.05524725  0.034260198
#> [23,] 0.21457663  0.21194724  0.130905342 0.129793731  0.14360112  0.082116123
#> [24,] 0.05629546 -0.05236511  0.008942650 0.005118801 -0.01755530 -0.002979762
#> [25,] 0.14347337  0.10585020  0.071357164 0.071669236  0.05746359  0.024800411
#>                V5         V50         V51         V52         V53         V54
#>  [1,]  0.01206910 0.015991030 0.043268998 0.017016988 0.011524447 0.032625059
#>  [2,]  0.12736793 0.017315590 0.029614505 0.017274767 0.027641011 0.012068251
#>  [3,]  0.11917682 0.017942958 0.014518545 0.015454964 0.013969330 0.016442714
#>  [4,]  0.05652675 0.023038002 0.014297406 0.009522664 0.008711427 0.013479759
#>  [5,]  0.14468125 0.027371842 0.021884636 0.034155612 0.011194022 0.008970949
#>  [6,]  0.05534177 0.014942679 0.018821376 0.013324754 0.006728194 0.004860674
#>  [7,]  0.08366668 0.024124503 0.017938291 0.009059038 0.007678544 0.010126067
#>  [8,]  0.07528540 0.021331521 0.020897497 0.012352299 0.008858049 0.007231026
#>  [9,]  0.08951283 0.034174426 0.029028360 0.023630713 0.012714658 0.012427373
#> [10,]  0.06450000 0.007700000 0.013700000 0.007100000 0.008200000 0.023200000
#> [11,]  0.06432136 0.013049659 0.017779271 0.015915515 0.005188013 0.010775699
#> [12,]  0.08447698 0.013171542 0.015259784 0.015976292 0.010087594 0.006717652
#> [13,]  0.12520776 0.002160564 0.001075353 0.005808642 0.011589947 0.016961155
#> [14,]  0.05377954 0.007168041 0.004269548 0.003370960 0.011235135 0.008929477
#> [15,]  0.02629725 0.035214004 0.015934143 0.010163107 0.004767347 0.010515537
#> [16,]  0.02319363 0.026810377 0.021325411 0.010828839 0.006781172 0.013339536
#> [17,]  0.01931836 0.019502287 0.020493404 0.017947710 0.017023026 0.005635968
#> [18,]  0.16957082 0.015969741 0.019899930 0.020482999 0.024596916 0.012919456
#> [19,]  0.05621281 0.007614514 0.011446037 0.005901905 0.006746234 0.006431071
#> [20,]  0.05664111 0.017686653 0.010588053 0.010265623 0.014139003 0.011238067
#> [21,]  0.05406136 0.013322060 0.009974836 0.009956244 0.004853356 0.003848797
#> [22,]  0.06203881 0.009226784 0.008005314 0.007359056 0.013028297 0.011324850
#> [23,]  0.10399819 0.015924682 0.019436245 0.012193634 0.006993027 0.018246817
#> [24,] -0.01281347 0.004413753 0.010204080 0.004093009 0.001376997 0.002880439
#> [25,]  0.03574375 0.012564474 0.009103700 0.011197179 0.008341983 0.008302421
#>               V55         V56         V57          V58         V59         V6
#>  [1,] 0.004317512 0.008568364 0.009138295  0.016151029 0.012555966 0.14883221
#>  [2,] 0.016170768 0.016519519 0.015095962  0.017306261 0.013605813 0.11593378
#>  [3,] 0.006971782 0.011757479 0.006661991  0.007924440 0.004184990 0.14831511
#>  [4,] 0.004893370 0.008188357 0.005811464  0.007656708 0.011835652 0.12159137
#>  [5,] 0.016073519 0.011905112 0.003364079  0.010322234 0.023552177 0.04745851
#>  [6,] 0.005094627 0.004389340 0.003463789  0.004487750 0.003964593 0.05810339
#>  [7,] 0.004971808 0.004717054 0.003773184  0.003841394 0.004893237 0.07252981
#>  [8,] 0.002356187 0.002764937 0.005296507  0.001839364 0.002491333 0.07971641
#>  [9,] 0.011941674 0.011235681 0.009540851  0.009559133 0.011046072 0.10110638
#> [10,] 0.019800000 0.007400000 0.003500000  0.010000000 0.004800000 0.04720000
#> [11,] 0.003619031 0.003017106 0.004653337  0.005498006 0.003601448 0.10353673
#> [12,] 0.007325120 0.005968536 0.005599306  0.008087467 0.006215093 0.09652184
#> [13,] 0.004870515 0.012046565 0.011391776  0.015423842 0.014102507 0.15042511
#> [14,] 0.006047623 0.005095943 0.003338727  0.003717079 0.003127005 0.09802789
#> [15,] 0.004481660 0.005120425 0.012006961  0.006500807 0.007129331 0.05581727
#> [16,] 0.006660079 0.005483040 0.003687888  0.006053132 0.005322495 0.03728474
#> [17,] 0.004497013 0.006589141 0.004312906  0.005157944 0.020527974 0.05735263
#> [18,] 0.006961075 0.013677859 0.009204184  0.013990376 0.003751432 0.17144038
#> [19,] 0.004501110 0.006098985 0.013339006  0.010152099 0.005641079 0.08184301
#> [20,] 0.005544571 0.010398268 0.015774830  0.007305702 0.007161103 0.07873573
#> [21,] 0.005499900 0.005370756 0.007332027  0.007746489 0.003773510 0.09171293
#> [22,] 0.011541311 0.006937019 0.009339988  0.008262808 0.010732100 0.10810606
#> [23,] 0.013423161 0.005881528 0.007452691  0.009463982 0.012979142 0.14683295
#> [24,] 0.006343299 0.003626620 0.003612704 -0.000642926 0.003934587 0.07733747
#> [25,] 0.006883821 0.006149076 0.005310809  0.007150974 0.005494681 0.06606645
#>                V60         V7         V8         V9
#>  [1,] 0.0135912685 0.18421975 0.24807799 0.28496459
#>  [2,] 0.0150917385 0.11850952 0.07601851 0.21142366
#>  [3,] 0.0039377570 0.12673775 0.19434580 0.27020141
#>  [4,] 0.0105320580 0.15466926 0.16289894 0.21085568
#>  [5,] 0.0127418987 0.07754158 0.05519508 0.12880211
#>  [6,] 0.0050400161 0.09448264 0.12114967 0.18287653
#>  [7,] 0.0068712684 0.09351126 0.11938990 0.16169883
#>  [8,] 0.0031742292 0.07008253 0.11439355 0.18090199
#>  [9,] 0.0094467364 0.11147904 0.15828183 0.26763871
#> [10,] 0.0019000000 0.10560000 0.13880000 0.05980000
#> [11,] 0.0060946836 0.06418724 0.05380586 0.16696050
#> [12,] 0.0048920757 0.11382135 0.08767990 0.12326211
#> [13,] 0.0063251007 0.16695032 0.07741654 0.08183307
#> [14,] 0.0041190527 0.09693865 0.07595931 0.14735160
#> [15,] 0.0037946233 0.13806864 0.21285601 0.21866418
#> [16,] 0.0104103854 0.08895951 0.09776983 0.04471065
#> [17,] 0.0109919909 0.03176997 0.10622264 0.17927951
#> [18,] 0.0103858750 0.07299231 0.13579168 0.20213007
#> [19,] 0.0062734194 0.06651539 0.09158576 0.06154165
#> [20,] 0.0098398478 0.09641157 0.08661166 0.05126386
#> [21,] 0.0056928395 0.08879128 0.09267021 0.10179736
#> [22,] 0.0053930505 0.09568569 0.08055875 0.08237898
#> [23,] 0.0085167333 0.20999796 0.21917983 0.28052157
#> [24,] 0.0008875974 0.07575344 0.05830477 0.06581937
#> [25,] 0.0048537903 0.08686261 0.08975590 0.09293947
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478