Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                  V1        V10        V11        V12        V13        V14
#>  [1,]  0.0631735925 0.12589272 0.20610185 0.23541720 0.28436088 0.29615161
#>  [2,]  0.0543162953 0.54300469 0.40938126 0.15032148 0.23904119 0.34358672
#>  [3,]  0.0367270154 0.31754872 0.37039129 0.37136795 0.29072129 0.14940523
#>  [4,]  0.0107429000 0.23277539 0.12928158 0.07832644 0.27397412 0.36641713
#>  [5,]  0.0334574655 0.33323318 0.29010090 0.29911880 0.36593496 0.34404576
#>  [6,]  0.0901348700 0.30553951 0.42668939 0.43825057 0.62998417 0.81247550
#>  [7,]  0.0257680281 0.16716375 0.20350216 0.24624404 0.31096622 0.38207879
#>  [8,]  0.0120826508 0.31615117 0.32228807 0.33961413 0.25014797 0.18214398
#>  [9,]  0.0325907753 0.61353164 0.63910590 0.46237757 0.30306226 0.19730079
#> [10,]  0.0124774226 0.20065642 0.27324289 0.30495032 0.26002056 0.24617365
#> [11,]  0.0325763297 0.22991123 0.32320097 0.38553069 0.41209052 0.44012312
#> [12,]  0.0041650588 0.12325532 0.20243460 0.23775200 0.21266939 0.14027107
#> [13,]  0.0235361989 0.31245846 0.24057265 0.28216168 0.24893008 0.19730912
#> [14,]  0.0061950619 0.09540965 0.09781634 0.05791480 0.07420498 0.08586566
#> [15,]  0.0149084249 0.09549683 0.10935429 0.09561971 0.11036895 0.12174335
#> [16,]  0.0121656727 0.17739608 0.13020586 0.07084671 0.07105770 0.08044318
#> [17,]  0.0236844553 0.09301850 0.07223174 0.18165475 0.16370982 0.11454004
#> [18,] -0.0003754364 0.10582802 0.04571895 0.08640686 0.22845476 0.18585893
#> [19,]  0.0254426541 0.15841619 0.16853465 0.16467307 0.17226425 0.13136765
#> [20,]  0.0129851951 0.13322645 0.12847621 0.08783436 0.06006233 0.10638433
#> [21,]  0.0206753612 0.07573702 0.13848988 0.25632157 0.21089347 0.21864648
#> [22,]  0.0166911537 0.39492897 0.30871572 0.15965588 0.29188824 0.51738013
#> [23,]  0.0392447874 0.16649441 0.21298886 0.31515741 0.26378508 0.18408449
#> [24,]  0.0289237088 0.30980881 0.26432895 0.07851121 0.20542359 0.48823632
#> [25,]  0.0311019566 0.24891360 0.32280706 0.37777044 0.39819918 0.48886597
#>              V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.34378068 0.3519926 0.3689907 0.2242849 0.2451498 0.091044831 0.3994120
#>  [2,] 0.45941134 0.4610162 0.1376358 0.1127737 0.3161128 0.087030999 0.9046584
#>  [3,] 0.16165498 0.2021197 0.2144204 0.2790987 0.3325171 0.051877806 0.3471130
#>  [4,] 0.50215217 0.6584790 0.6962099 0.7613353 0.8670216 0.006059588 0.8798200
#>  [5,] 0.26101785 0.2876290 0.2705144 0.2206059 0.3761413 0.030178354 0.5083133
#>  [6,] 0.90893261 0.9406040 0.8682876 0.8978731 0.9144885 0.090852567 0.8916050
#>  [7,] 0.53710836 0.6980775 0.8085330 0.8686034 0.9161822 0.026809528 0.9102553
#>  [8,] 0.11921985 0.2382011 0.3046802 0.3595396 0.2546541 0.038643493 0.1436247
#>  [9,] 0.23211654 0.2957857 0.2640897 0.3294985 0.4387536 0.061561054 0.5687786
#> [10,] 0.24211974 0.1879186 0.1691259 0.2494459 0.3626811 0.026597626 0.4104202
#> [11,] 0.46917850 0.4962130 0.5447265 0.6247292 0.7466841 0.037313360 0.8205414
#> [12,] 0.02624773 0.1547367 0.3068767 0.4276545 0.5007955 0.017142235 0.6093656
#> [13,] 0.13034584 0.1534454 0.1253130 0.1812502 0.1698681 0.027008744 0.1406595
#> [14,] 0.17036197 0.2036361 0.2135598 0.2523833 0.2355458 0.009555701 0.2222802
#> [15,] 0.11105890 0.1688115 0.2319985 0.2713634 0.3137961 0.015850718 0.3734656
#> [16,] 0.11515714 0.2257052 0.2831260 0.2809840 0.3921438 0.002073950 0.5762488
#> [17,] 0.14132900 0.1868107 0.1699439 0.1419803 0.1482640 0.030942193 0.1856734
#> [18,] 0.10208290 0.2430235 0.2893045 0.3542462 0.5760855 0.020301884 0.6830984
#> [19,] 0.13996319 0.1379916 0.2159028 0.3439769 0.3911452 0.031558369 0.4879692
#> [20,] 0.08870161 0.1038693 0.2299620 0.2891431 0.3579039 0.001790866 0.5264472
#> [21,] 0.22829122 0.3837769 0.6712181 0.5899170 0.3493551 0.036523082 0.3064345
#> [22,] 0.65175928 0.9198919 0.9404310 0.9059220 0.6790629 0.047853804 0.5725903
#> [23,] 0.23362857 0.3199438 0.3869637 0.4906688 0.4925522 0.045910701 0.3944441
#> [24,] 0.62396523 0.7090012 0.7849226 0.6694301 0.4519683 0.088569518 0.3614296
#> [25,] 0.66482374 0.7780639 0.9484559 0.9657074 0.9043716 0.037431323 0.8333447
#>              V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.48925696 0.5592992 0.5935058 0.5573110 0.5545169 0.5486799 0.6389249
#>  [2,] 0.92960021 0.3734421 0.1746586 0.1753154 0.1589893 0.3855929 0.2973702
#>  [3,] 0.47856143 0.7112350 0.8331950 0.8822205 0.9137895 0.9624192 0.9345975
#>  [4,] 0.96053718 0.9901174 0.8828835 0.5987001 0.4054654 0.4450847 0.6552700
#>  [5,] 0.49770592 0.4580895 0.5092433 0.5121113 0.5845402 0.6952855 0.6971582
#>  [6,] 0.79386674 0.7366884 0.7643411 0.7186093 0.5897856 0.6072527 0.5558077
#>  [7,] 0.84241618 0.8578189 0.7827372 0.5382273 0.3096774 0.1736150 0.1505542
#>  [8,] 0.07527773 0.1119933 0.3045917 0.4322776 0.5743297 0.7799696 0.9264494
#>  [9,] 0.69659779 0.8435798 0.9410525 0.9816595 0.7596652 0.6370944 0.5291642
#> [10,] 0.43554265 0.3636175 0.3788011 0.5019162 0.6216279 0.7626479 0.9035975
#> [11,] 0.87314551 0.9716773 0.9476956 0.9181477 0.8461386 0.7770613 0.7889943
#> [12,] 0.71341307 0.8138281 0.8431333 0.7842785 0.8407600 0.9772404 1.0155028
#> [13,] 0.07223913 0.2160716 0.4335709 0.5948195 0.6903255 0.7784290 0.8755679
#> [14,] 0.29341918 0.3403718 0.4847809 0.7284639 0.8864871 0.8710558 0.7841452
#> [15,] 0.45679724 0.5584571 0.6381713 0.6724563 0.6602418 0.7147956 0.7778783
#> [16,] 0.67137056 0.8833731 1.0229666 0.8832358 0.8173958 0.9007007 0.9233005
#> [17,] 0.22547240 0.1561759 0.2237025 0.2078149 0.1013956 0.2729981 0.5409051
#> [18,] 0.73404211 0.7695594 0.4287731 0.3147830 0.6791046 0.8231958 0.9425648
#> [19,] 0.61197416 0.8087209 0.8788198 0.8251598 0.6853890 0.6445337 0.5779635
#> [20,] 0.64752278 0.7871293 0.9243815 0.8962356 0.9016341 0.9836563 0.9881840
#> [21,] 0.36354064 0.4703062 0.5845392 0.7471132 0.8419356 0.8122720 0.7754267
#> [22,] 0.61247185 0.5858098 0.6944654 0.7869942 0.8133502 0.7704611 0.7656771
#> [23,] 0.27169141 0.2195700 0.2908013 0.2359711 0.2473179 0.3335355 0.3067611
#> [24,] 0.53617835 0.7556814 0.8408411 0.9341518 0.8058486 0.6929059 0.6859430
#> [25,] 0.78729669 0.8328180 0.6546905 0.5841649 0.6962006 0.7622191 0.5943641
#>             V28       V29            V3       V30       V31       V32       V33
#>  [1,] 0.7525511 0.7211138  0.1083484255 0.7334692 0.7854195 0.7378727 0.5478503
#>  [2,] 0.6011594 0.5043647  0.0903293357 0.7531727 0.3257836 0.3757099 0.6258372
#>  [3,] 0.8746238 0.7631611  0.0527067382 0.5949284 0.5820344 0.5297640 0.4145584
#>  [4,] 0.8514668 0.8647882  0.0267748017 0.6477496 0.3167473 0.2986783 0.3259965
#>  [5,] 0.7677565 0.7618950  0.0547631914 0.8120274 0.8026656 0.7064380 0.5440591
#>  [6,] 0.5366951 0.5398070  0.0680307229 0.3863250 0.3427512 0.2730689 0.1669996
#>  [7,] 0.2538959 0.3844378  0.0219993835 0.4680377 0.3980377 0.4293040 0.4428133
#>  [8,] 0.9411954 0.9499639  0.0259299208 1.0231798 0.9488211 0.8305625 0.7263913
#>  [9,] 0.4408600 0.3958467  0.0634881877 0.3846823 0.4179231 0.2423369 0.4909472
#> [10,] 0.9578570 0.8995881  0.0326464187 0.6961623 0.5790887 0.5015677 0.4486384
#> [11,] 0.7008391 0.5810369  0.0601470150 0.4513025 0.3234550 0.2429882 0.2349886
#> [12,] 0.9842344 0.8739133  0.0239981719 0.6290459 0.2478064 0.2461421 0.3085035
#> [13,] 0.9162186 0.8717999  0.0197653483 0.8977659 0.9120561 0.8571174 0.7298301
#> [14,] 0.7067521 0.5357653  0.0043381138 0.3716858 0.3023824 0.2546640 0.2283400
#> [15,] 0.8581855 0.9418151  0.0157950443 0.8759743 0.7599887 0.6667704 0.6303625
#> [16,] 0.9128038 0.8295121 -0.0008236045 0.7115591 0.5928042 0.5099484 0.4967342
#> [17,] 0.7122817 0.7323053  0.0204932298 0.7186399 0.7178090 0.5513201 0.6466000
#> [18,] 0.9010796 0.6728121  0.0019403825 0.4031292 0.3487850 0.2957592 0.3009756
#> [19,] 0.5730920 0.5987713  0.0487463584 0.5664266 0.5226997 0.6155650 0.7479277
#> [20,] 0.9228186 0.8416114  0.0079852100 0.7211831 0.6170113 0.5839583 0.5641321
#> [21,] 0.7912246 0.8708720  0.0424373774 0.8278356 0.7946584 0.7267306 0.6762677
#> [22,] 0.5751624 0.2498516  0.0526813549 0.2316037 0.5431091 0.6506190 0.3499140
#> [23,] 0.4127179 0.5807450  0.0681919761 0.6372022 0.7069384 0.7624951 0.6727543
#> [24,] 0.5332927 0.4061794  0.1173756022 0.4593880 0.4282343 0.1866157 0.1769976
#> [25,] 0.5257039 0.7135792  0.0498791849 0.7578549 0.5790457 0.3527886 0.3632393
#>              V34       V35        V36       V37       V38       V39          V4
#>  [1,] 0.53575577 0.6053672 0.40725983 0.2368207 0.3542546 0.3340689 0.112163888
#>  [2,] 0.80852384 0.8873967 0.91046400 0.8567801 0.9358547 0.7088222 0.113538766
#>  [3,] 0.23285639 0.1354236 0.09603337 0.1568851 0.1386392 0.1489529 0.061529919
#>  [4,] 0.20119001 0.1382375 0.09510979 0.0894435 0.1074599 0.1343712 0.056647086
#>  [5,] 0.36835628 0.1739225 0.15090847 0.2150582 0.3299342 0.4938715 0.046164143
#>  [6,] 0.21185070 0.2280286 0.13770132 0.1642191 0.1562973 0.2035934 0.039193241
#>  [7,] 0.43700244 0.3443046 0.23867169 0.1934224 0.3174897 0.2734428 0.049429697
#>  [8,] 0.68095606 0.7356458 0.78763773 0.7715394 0.6266486 0.5128412 0.069890625
#>  [9,] 0.64873875 0.5869600 0.44099677 0.4037064 0.3408435 0.4043032 0.066206193
#> [10,] 0.36842299 0.1958192 0.11698768 0.1950622 0.2005892 0.2130939 0.052917545
#> [11,] 0.18869458 0.2271007 0.33545407 0.3533420 0.3514823 0.3608312 0.077420676
#> [12,] 0.46173049 0.7461391 0.90391210 0.8361863 0.6525547 0.4874220 0.024543127
#> [13,] 0.63809311 0.6352791 0.68604536 0.7048614 0.5951548 0.5417581 0.029321501
#> [14,] 0.25776430 0.2830527 0.26624659 0.2149386 0.1829912 0.2122743 0.008172476
#> [15,] 0.63098776 0.6296533 0.56922467 0.4712665 0.3329278 0.2363096 0.024788459
#> [16,] 0.56169169 0.6090920 0.54837448 0.5219005 0.3585520 0.1982181 0.028071100
#> [17,] 0.82693091 0.8838303 0.87466010 0.7138148 0.6098655 0.5586231 0.020394423
#> [18,] 0.26865673 0.3353445 0.32443001 0.4682909 0.4820335 0.3907019 0.010240967
#> [19,] 0.78875820 0.7436991 0.65098292 0.5205922 0.3470311 0.3340115 0.051590766
#> [20,] 0.59117732 0.6044438 0.56098362 0.4962708 0.3878964 0.1871725 0.042005622
#> [21,] 0.80478408 0.8958831 0.94551427 0.8894598 0.7007008 0.5495502 0.041284563
#> [22,] 0.49707069 0.8252285 0.71444834 0.4950934 0.4859062 0.4935679 0.060139659
#> [23,] 0.58719027 0.6136233 0.65650805 0.7989764 0.9775686 0.9525698 0.024128661
#> [24,] 0.07742185 0.3818220 0.65696289 0.6626628 0.3216652 0.1426821 0.102219140
#> [25,] 0.41522807 0.2479028 0.13199417 0.2317501 0.1965747 0.1235159 0.028633287
#>              V40        V41        V42        V43        V44          V45
#>  [1,] 0.19174662 0.55594069 0.65987080 0.45612593 0.42281019  0.571194050
#>  [2,] 0.46192810 0.41745762 0.36034019 0.23289101 0.24937625  0.374444561
#>  [3,] 0.13043015 0.13950188 0.12515435 0.19501089 0.20061874  0.167061671
#>  [4,] 0.19633445 0.24823151 0.24774864 0.13697575 0.20318891  0.217474509
#>  [5,] 0.51266354 0.51172186 0.49195166 0.47660549 0.45695994  0.590054859
#>  [6,] 0.24297043 0.14117238 0.15681495 0.18423510 0.20948410  0.120696365
#>  [7,] 0.18338102 0.21641107 0.21789366 0.18450911 0.16246627  0.124202305
#>  [8,] 0.51986760 0.45225163 0.39580805 0.47464330 0.54308944  0.456379203
#>  [9,] 0.41005123 0.28419103 0.34508917 0.34857956 0.20056822  0.157817833
#> [10,] 0.26225428 0.32420872 0.35342286 0.22598219 0.12521587  0.160208804
#> [11,] 0.33985981 0.35529509 0.38918286 0.33719020 0.24926920  0.202024688
#> [12,] 0.33785858 0.30101971 0.23151514 0.33108044 0.37836223  0.336025506
#> [13,] 0.57615926 0.54293948 0.47743470 0.48661498 0.56002027  0.551237308
#> [14,] 0.22653047 0.23523117 0.19030136 0.15387399 0.13500802  0.117484789
#> [15,] 0.21018738 0.18612856 0.15640965 0.18736098 0.18444722  0.136361332
#> [16,] 0.08173666 0.09195917 0.02787654 0.08852159 0.03834574 -0.007813287
#> [17,] 0.61936916 0.52337966 0.46465684 0.32794454 0.19012839  0.151125100
#> [18,] 0.24284908 0.16877461 0.11969382 0.05997830 0.08264009  0.083867433
#> [19,] 0.44521677 0.43543168 0.35380893 0.37676687 0.30382162  0.186422806
#> [20,] 0.16833894 0.16799693 0.06623140 0.05100904 0.03283818  0.046236569
#> [21,] 0.40223337 0.29624677 0.23390679 0.21753489 0.19424078  0.209532178
#> [22,] 0.36626510 0.14493524 0.16190500 0.26560761 0.28111008  0.191974878
#> [23,] 0.82546853 0.70856802 0.61406377 0.44763580 0.34036824  0.289124764
#> [24,] 0.36935247 0.43139515 0.36412205 0.22683096 0.18469276  0.201535070
#> [25,] 0.21948967 0.20934637 0.22096929 0.14418335 0.11918024  0.195559282
#>              V46        V47        V48         V49          V5         V50
#>  [1,] 0.56037928 0.34976849 0.16771318 0.095487400 0.074563742 0.039400370
#>  [2,] 0.26893650 0.14876983 0.17100486 0.075410343 0.068979357 0.019102535
#>  [3,] 0.15167285 0.15190758 0.13014223 0.077803264 0.070723846 0.022318613
#>  [4,] 0.20707668 0.12856414 0.04445977 0.036484554 0.091117291 0.009114418
#>  [5,] 0.49063980 0.34227739 0.25393494 0.139130275 0.095423029 0.047661521
#>  [6,] 0.09103188 0.10701204 0.10386118 0.077201890 0.060698649 0.019661573
#>  [7,] 0.05348615 0.05030145 0.03875753 0.030444269 0.076909899 0.010682944
#>  [8,] 0.26438863 0.14642513 0.14695148 0.097800753 0.105277006 0.026786439
#>  [9,] 0.05896634 0.11241960 0.16290426 0.082512437 0.029034588 0.019415008
#> [10,] 0.19040731 0.19435613 0.14934960 0.062594618 0.080096296 0.017173524
#> [11,] 0.14278048 0.10853411 0.08628580 0.053839031 0.134821027 0.019352151
#> [12,] 0.25552658 0.11889018 0.04862682 0.039491893 0.038265583 0.011910603
#> [13,] 0.43495637 0.34189855 0.24894633 0.128021442 0.058792204 0.038769017
#> [14,] 0.11170053 0.03959767 0.03322239 0.024456262 0.023122209 0.009081342
#> [15,] 0.09815142 0.05778736 0.04525289 0.022069047 0.042204136 0.012232290
#> [16,] 0.03070178 0.03293556 0.03381167 0.018371100 0.059415601 0.008227409
#> [17,] 0.09539047 0.10528658 0.08831540 0.045279029 0.035827115 0.020641112
#> [18,] 0.05269379 0.04851814 0.03295065 0.003452262 0.001555257 0.015485531
#> [19,] 0.11565814 0.06651835 0.04684195 0.024442628 0.071674062 0.018576380
#> [20,] 0.01852895 0.01283095 0.01795153 0.006647986 0.054512348 0.010409533
#> [21,] 0.18634927 0.11266406 0.04508835 0.022394585 0.042967509 0.029314407
#> [22,] 0.07669994 0.06130602 0.04221025 0.022819717 0.118394476 0.024878593
#> [23,] 0.27438426 0.23024377 0.16305900 0.088978912 0.032808678 0.036650423
#> [24,] 0.15548366 0.15898167 0.15838628 0.098304676 0.083769687 0.019472915
#> [25,] 0.21958642 0.18640598 0.11780325 0.058270707 0.067956648 0.029280888
#>               V51         V52         V53          V54          V55         V56
#>  [1,] 0.045655542 0.040516038 0.021547637 0.0225914575 0.0237008576 0.011650050
#>  [2,] 0.033259258 0.027753253 0.017102613 0.0260977055 0.0173863526 0.022622154
#>  [3,] 0.018533396 0.015864606 0.006611141 0.0109431556 0.0045573625 0.006815823
#>  [4,] 0.003476931 0.005977786 0.002960104 0.0005143765 0.0004491949 0.001657067
#>  [5,] 0.017281123 0.014733544 0.015394675 0.0104138554 0.0121859588 0.011140113
#>  [6,] 0.026543293 0.014926484 0.017098192 0.0160724818 0.0215125014 0.008708255
#>  [7,] 0.013856860 0.013933248 0.004993999 0.0125257914 0.0118259456 0.006840736
#>  [8,] 0.010952195 0.018042541 0.012110662 0.0110003093 0.0018780529 0.007988227
#>  [9,] 0.016624026 0.014952039 0.015120014 0.0185869007 0.0120218633 0.006744586
#> [10,] 0.018759651 0.015201685 0.009335342 0.0076358569 0.0061878403 0.004162764
#> [11,] 0.022560966 0.020946738 0.018324291 0.0230529645 0.0074710828 0.010975808
#> [12,] 0.003993606 0.010947754 0.006921976 0.0053212277 0.0028229226 0.001593888
#> [13,] 0.023949419 0.016702149 0.007753977 0.0075711205 0.0079389465 0.008532505
#> [14,] 0.002041661 0.008784370 0.004545612 0.0071960188 0.0038921756 0.001499753
#> [15,] 0.005808698 0.005879131 0.004700088 0.0067086696 0.0055757360 0.004249189
#> [16,] 0.000478292 0.003356310 0.008945861 0.0094924022 0.0093371780 0.002177612
#> [17,] 0.013811083 0.010406476 0.017443162 0.0140651002 0.0107868541 0.007600002
#> [18,] 0.015407474 0.010704241 0.022963884 0.0245402791 0.0143354807 0.012990424
#> [19,] 0.009185627 0.011269933 0.014935493 0.0124866054 0.0053105477 0.012571414
#> [20,] 0.007765400 0.007387192 0.005259711 0.0013111565 0.0006119420 0.004557961
#> [21,] 0.011591616 0.010012668 0.007103297 0.0121055342 0.0122194018 0.004280960
#> [22,] 0.003666591 0.019878366 0.006575830 0.0185287891 0.0117907608 0.002851887
#> [23,] 0.021121750 0.014627149 0.014068840 0.0080860076 0.0083303514 0.005672176
#> [24,] 0.021235364 0.011119226 0.007979842 0.0143345194 0.0102709953 0.006440603
#> [25,] 0.010261346 0.005739467 0.006621268 0.0091878182 0.0106511840 0.009391608
#>                V57         V58         V59           V6         V60         V7
#>  [1,] 0.0151815920 0.014199583 0.022078143  0.108596174 0.013812381 0.06395084
#>  [2,] 0.0152707137 0.016768670 0.019491042  0.133062767 0.015989634 0.25298548
#>  [3,] 0.0062599799 0.008391009 0.010343376  0.136070128 0.008266547 0.15771805
#>  [4,] 0.0006177664 0.006358631 0.002000087  0.126836128 0.005258859 0.15757878
#>  [5,] 0.0111884889 0.011743240 0.009169155  0.115941041 0.004661977 0.12207321
#>  [6,] 0.0019640644 0.017911793 0.010805432  0.134338287 0.010488792 0.18262024
#>  [7,] 0.0033706649 0.009576680 0.005969027  0.121802030 0.002104565 0.14699251
#>  [8,] 0.0070836398 0.005616083 0.001370260  0.107255954 0.003177444 0.15683707
#>  [9,] 0.0040232082 0.014192066 0.007280478  0.047147948 0.003651068 0.09544832
#> [10,] 0.0040175076 0.006965974 0.005866572  0.091405216 0.002084385 0.12054248
#> [11,] 0.0108933895 0.011709088 0.012903125  0.164017378 0.008363508 0.13642958
#> [12,] 0.0003521493 0.000613994 0.002545922  0.062199703 0.004315724 0.06930747
#> [13,] 0.0072767010 0.003877471 0.004319182  0.081670576 0.002520648 0.11368173
#> [14,] 0.0023395553 0.003097833 0.002165417  0.031881857 0.002905617 0.05085055
#> [15,] 0.0067988819 0.005386370 0.004676803  0.072569116 0.003176994 0.06845345
#> [16,] 0.0033158755 0.001243285 0.001948712  0.102725501 0.002878698 0.11011533
#> [17,] 0.0074113721 0.007718680 0.010660980  0.062492796 0.006270455 0.06060716
#> [18,] 0.0054242776 0.012101031 0.002359327 -0.003797821 0.004305332 0.01159861
#> [19,] 0.0170527388 0.007747158 0.006825586  0.124468200 0.009396777 0.13371547
#> [20,] 0.0038487351 0.001887637 0.001693318  0.082762660 0.001719073 0.08961895
#> [21,] 0.0107668903 0.006056694 0.009236877  0.090872336 0.008388179 0.14876185
#> [22,] 0.0035201691 0.005378140 0.010096226  0.239885189 0.006418705 0.19928925
#> [23,] 0.0020191156 0.012858563 0.007175206  0.063364904 0.007527198 0.10955484
#> [24,] 0.0095626339 0.009486788 0.010705034  0.105711231 0.008018124 0.15160389
#> [25,] 0.0073773709 0.005540162 0.003455891  0.119419354 0.006601731 0.12879152
#>                 V8         V9
#>  [1,] 0.1199690411 0.17379504
#>  [2,] 0.4495776752 0.64283637
#>  [3,] 0.1576191084 0.22004251
#>  [4,] 0.1949022425 0.24894376
#>  [5,] 0.1271158639 0.24691292
#>  [6,] 0.1588034486 0.20155250
#>  [7,] 0.1182099675 0.13301305
#>  [8,] 0.2133470566 0.32794986
#>  [9,] 0.2165814220 0.38618517
#> [10,] 0.1046661564 0.15876101
#> [11,] 0.1638091564 0.17976162
#> [12,] 0.0001964127 0.06824489
#> [13,] 0.1108809889 0.25341014
#> [14,] 0.0618267100 0.07451552
#> [15,] 0.0762479906 0.08863011
#> [16,] 0.0878956584 0.15670896
#> [17,] 0.0795732248 0.06555454
#> [18,] 0.0375523976 0.09260272
#> [19,] 0.1142340973 0.09218652
#> [20,] 0.0902810569 0.16363723
#> [21,] 0.1379352522 0.08443743
#> [22,] 0.1906500893 0.30062103
#> [23,] 0.1463375352 0.14404310
#> [24,] 0.2005253206 0.24731855
#> [25,] 0.1345792432 0.22571688
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841