Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13       V14
#>  [1,] 0.024279617  0.07809006 0.22958950 0.42755108 0.44686482 0.5321183
#>  [2,] 0.058142379  0.35600315 0.52052417 0.51733599 0.54100945 0.5579348
#>  [3,] 0.039346253  0.22658973 0.25311169 0.19380729 0.19921340 0.1894608
#>  [4,] 0.027463376  0.57112226 0.64753449 0.56696562 0.52030314 0.3432110
#>  [5,] 0.028636297  0.25312986 0.30260784 0.32418918 0.28761765 0.2189403
#>  [6,] 0.043601286  0.06793912 0.13892298 0.19169746 0.24784807 0.3185529
#>  [7,] 0.029628146  0.35414098 0.44573678 0.43943726 0.37113285 0.3109161
#>  [8,] 0.048150647  0.15591645 0.19330701 0.23407989 0.25020667 0.2642804
#>  [9,] 0.025895439  0.34916463 0.39418028 0.42513868 0.49166468 0.4364553
#> [10,] 0.050141056 -0.02422749 0.10528931 0.23470679 0.32751695 0.4570607
#> [11,] 0.020055660  0.29101898 0.34362473 0.36776874 0.37959869 0.4143984
#> [12,] 0.024341329  0.41836423 0.39597600 0.44851840 0.46477935 0.5087596
#> [13,] 0.003811236  0.22169384 0.12761368 0.05457683 0.22087258 0.3278064
#> [14,] 0.022972867  0.62029621 0.63734473 0.72490290 0.56675037 0.5423418
#> [15,] 0.010610811  0.06403674 0.05616788 0.06201711 0.07070762 0.1136089
#> [16,] 0.015329119  0.08634401 0.06260904 0.09456405 0.25937835 0.2976928
#> [17,] 0.018975407  0.13967950 0.17415511 0.18013797 0.17053310 0.1613545
#> [18,] 0.009828002  0.09202151 0.10233970 0.11635917 0.14655753 0.2030859
#> [19,] 0.030185377  0.02759336 0.04157908 0.18745760 0.14595282 0.1220644
#> [20,] 0.021749283  0.24027401 0.26538451 0.14263265 0.24110701 0.4503846
#> [21,] 0.013931309  0.12905311 0.13513192 0.21520867 0.28100265 0.1531722
#> [22,] 0.021319345  0.07189116 0.11026560 0.19518272 0.16966146 0.1856228
#> [23,] 0.006322993  0.24629744 0.15834049 0.08798477 0.24017407 0.3247742
#> [24,] 0.020888767  0.11437037 0.20225082 0.26005889 0.32723907 0.3201109
#> [25,] 0.036172074  0.26754467 0.23854201 0.10296507 0.17434170 0.4742894
#>             V15       V16        V17       V18       V19          V2       V20
#>  [1,] 0.6880239 0.7046986 0.79766259 0.9384683 1.0019793 0.034761943 0.9886866
#>  [2,] 0.6602249 0.8337100 0.95280814 0.9325970 0.8138716 0.094763972 0.7495990
#>  [3,] 0.2040816 0.2457257 0.18839356 0.2577725 0.3453959 0.067383073 0.6596290
#>  [4,] 0.2035390 0.2708698 0.28490583 0.3327264 0.5847317 0.036073976 0.7439707
#>  [5,] 0.2057142 0.1906112 0.19368377 0.2402739 0.2906304 0.042330126 0.3283809
#>  [6,] 0.3970705 0.4286929 0.55669940 0.6742532 0.8383478 0.062134357 0.9083111
#>  [7,] 0.2925645 0.3512459 0.31897968 0.3138776 0.3900231 0.042253340 0.5413941
#>  [8,] 0.2929227 0.3543066 0.36169194 0.2587695 0.3325836 0.077202071 0.3152519
#>  [9,] 0.3124881 0.3498497 0.35896030 0.3292581 0.4686597 0.042299037 0.7164260
#> [10,] 0.5557679 0.6083252 0.73131213 0.8077367 0.8563498 0.053003878 0.8713812
#> [11,] 0.4248591 0.3939880 0.48230972 0.5673233 0.6790705 0.023304934 0.7132806
#> [12,] 0.4659079 0.5754664 0.80179639 0.8350589 0.8122593 0.115816470 0.8403880
#> [13,] 0.4346907 0.6240992 0.65519912 0.7386579 0.8812328 0.009068364 0.9336931
#> [14,] 0.6472436 0.6942753 0.69588778 0.7814909 0.9109735 0.056281802 0.8615109
#> [15,] 0.1199685 0.2104253 0.25791897 0.3460935 0.3542306 0.012562193 0.3619334
#> [16,] 0.2161513 0.2099539 0.13685363 0.3226250 0.6340711 0.030413170 0.8622005
#> [17,] 0.1819137 0.2029848 0.23193748 0.3020740 0.3358182 0.026911857 0.4612713
#> [18,] 0.2136845 0.2138126 0.13477325 0.1583887 0.1750888 0.011181425 0.2000843
#> [19,] 0.1101930 0.1268439 0.15245028 0.1335156 0.1686434 0.046373719 0.1022852
#> [20,] 0.5807776 0.7980959 0.94496520 0.9383648 0.7504491 0.026928620 0.5174223
#> [21,] 0.2025187 0.2666493 0.27748215 0.3423879 0.3506650 0.040730848 0.1433200
#> [22,] 0.2180037 0.4296767 0.73743622 0.6357314 0.3295885 0.025104906 0.2588442
#> [23,] 0.4370945 0.2688629 0.06376809 0.1423347 0.2391252 0.022695377 0.2641571
#> [24,] 0.4054946 0.4997584 0.46481122 0.4963555 0.5520087 0.023506231 0.5923371
#> [25,] 0.6656905 0.7120529 0.74303012 0.5920388 0.3843440 0.076469802 0.3271748
#>             V21        V22        V23        V24       V25       V26       V27
#>  [1,] 0.8281227 0.75537149 0.82351811 0.81840947 0.5245140 0.2870926 0.2859202
#>  [2,] 0.6791936 0.39001541 0.09459475 0.07837423 0.2639872 0.2988143 0.2579733
#>  [3,] 0.7174470 0.61283823 0.58080866 0.49299702 0.4906386 0.6554089 0.7354797
#>  [4,] 0.8854981 0.82339798 0.71764678 0.74107033 0.7194267 0.8121263 0.8885259
#>  [5,] 0.4219246 0.52891081 0.63349481 0.72430617 0.7740623 0.8117507 0.8689112
#>  [6,] 0.8994288 0.98388319 0.98763813 0.93105065 0.8568510 0.7413424 0.6965436
#>  [7,] 0.7323957 0.90284726 0.93966098 0.92922329 0.8798339 0.8207735 0.7598398
#>  [8,] 0.3662184 0.38800214 0.44322419 0.53540299 0.5281909 0.5407857 0.6406263
#>  [9,] 0.7435240 0.71091630 0.64700127 0.64973366 0.6916284 0.8644880 0.9668782
#> [10,] 0.8480655 0.84397639 0.90779510 0.95294292 0.8103603 0.5504246 0.3914708
#> [11,] 0.7637073 0.85395650 0.89555416 0.96413105 0.9578713 0.9415642 0.9686839
#> [12,] 0.7506144 0.81465872 1.01463845 1.09544774 0.9962588 0.9572467 0.9577518
#> [13,] 1.0035538 0.99933406 0.92692136 0.62052151 0.3544685 0.3675340 0.6163692
#> [14,] 0.7973292 0.68559047 0.43948060 0.37280691 0.5458552 0.2395941 0.5182380
#> [15,] 0.3958370 0.40185701 0.37680744 0.43201907 0.3859707 0.4429831 0.6149172
#> [16,] 0.8444523 0.82446827 0.53612199 0.40313545 0.5566672 0.7837942 0.9307373
#> [17,] 0.5468987 0.68149554 0.78275872 0.74827957 0.7593419 0.8103318 0.7581423
#> [18,] 0.2249548 0.25958079 0.42865843 0.60435240 0.7264365 0.8276132 0.8011422
#> [19,] 0.1549193 0.08293156 0.23830942 0.27879795 0.1738635 0.2782794 0.4730753
#> [20,] 0.5673050 0.69837955 0.75038330 0.79166240 0.7422346 0.6993481 0.6298079
#> [21,] 0.1897444 0.20696376 0.26131206 0.27871241 0.3102428 0.3456225 0.2307331
#> [22,] 0.4073479 0.57679967 0.64947082 0.70436986 0.8368027 0.9269798 0.9103986
#> [23,] 0.2539897 0.39412108 0.64231648 0.80788009 0.7013322 0.6821655 0.6923089
#> [24,] 0.5477671 0.45465938 0.57425475 0.73753947 0.7072090 0.7290957 0.7814267
#> [25,] 0.5683896 0.79270226 0.86642220 0.91813879 0.8178033 0.6775888 0.6620976
#>             V28        V29         V3       V30       V31       V32        V33
#>  [1,] 0.3584499 0.38999245 0.04285255 0.3126988 0.2863718 0.3598680 0.29730122
#>  [2,] 0.2262377 0.09573876 0.13749860 0.2322133 0.2302175 0.2272315 0.26772087
#>  [3,] 0.8167277 0.65879085 0.04946474 0.6155116 0.3425489 0.2560219 0.32699390
#>  [4,] 0.9777266 0.82022221 0.04951066 0.4969195 0.3362394 0.2983037 0.26373057
#>  [5,] 0.9032990 0.87112826 0.04449920 0.7290105 0.6608091 0.5126476 0.41650588
#>  [6,] 0.5550415 0.37677670 0.14370137 0.2865053 0.1781674 0.1132318 0.15182173
#>  [7,] 0.6550504 0.49551450 0.04206052 0.3711341 0.3187392 0.3210340 0.35528221
#>  [8,] 0.7475716 0.73984399 0.08216419 0.7566461 0.7335793 0.6448818 0.49283193
#>  [9,] 0.9149373 0.77872401 0.05879787 0.6267545 0.5683067 0.6829068 0.57130543
#> [10,] 0.3188954 0.35817776 0.07327429 0.2920230 0.2688963 0.2907573 0.38526609
#> [11,] 0.9469313 0.85826690 0.04349411 0.6213286 0.3853476 0.2968647 0.15148465
#> [12,] 0.7464314 0.66816061 0.13100125 0.5710193 0.2637575 0.1704885 0.07479905
#> [13,] 0.8526423 0.79138985 0.01638449 0.5691930 0.2755064 0.2913270 0.25519710
#> [14,] 0.8627298 0.60450014 0.11320968 0.8576459 0.8696234 0.5074166 0.16289468
#> [15,] 0.8222764 0.95033903 0.01147443 0.9036682 0.7566891 0.5761202 0.60744053
#> [16,] 0.8003076 0.57468716 0.02215078 0.4053923 0.3734567 0.2776653 0.29624928
#> [17,] 0.6982864 0.64164888 0.03035582 0.5887471 0.5408452 0.5375966 0.62459423
#> [18,] 0.7977764 0.78497976 0.01629950 0.6385419 0.4921434 0.3782841 0.41144097
#> [19,] 0.6490851 0.69081594 0.02721493 0.7008918 0.7842419 0.6501840 0.70643089
#> [20,] 0.4589247 0.37262973 0.05260032 0.4506933 0.3883463 0.3259349 0.25065345
#> [21,] 0.4797312 0.66709381 0.03983314 0.6633850 0.6237615 0.6367818 0.64485996
#> [22,] 0.8332445 0.84692604 0.03260693 0.8800266 0.8484768 0.7204172 0.68229405
#> [23,] 0.8522877 1.00168710 0.05121464 0.9627954 0.8999565 0.7572113 0.69029289
#> [24,] 0.8382844 0.76263675 0.02503760 0.6355190 0.5292459 0.4309894 0.33229443
#> [25,] 0.5752002 0.42195273 0.10703872 0.4646363 0.4536975 0.2288415 0.20046733
#>              V34       V35       V36        V37       V38       V39         V4
#>  [1,] 0.27670318 0.3163906 0.2918734 0.35142262 0.4714288 0.3635374 0.07365514
#>  [2,] 0.24209315 0.3092463 0.1431748 0.12559504 0.2066687 0.1965204 0.12541288
#>  [3,] 0.53772485 0.7498598 0.8308839 0.80629198 0.7505922 0.6454351 0.05295683
#>  [4,] 0.29446652 0.1621685 0.2925290 0.18955199 0.3254828 0.5880325 0.06360970
#>  [5,] 0.31764671 0.1925185 0.1251930 0.20560399 0.1527871 0.1839451 0.05511887
#>  [6,] 0.24235095 0.4275675 0.4722820 0.39909419 0.3838085 0.3791323 0.21138658
#>  [7,] 0.38309304 0.2803066 0.2949712 0.35554850 0.3443063 0.4099799 0.04738865
#>  [8,] 0.46894649 0.4650658 0.3535642 0.33834170 0.3859315 0.3388137 0.09416711
#>  [9,] 0.33145050 0.0852287 0.2505778 0.27622948 0.4060834 0.5915068 0.04913041
#> [10,] 0.43944023 0.4814547 0.4179279 0.36326633 0.4487047 0.4248348 0.07526389
#> [11,] 0.02715622 0.1076936 0.1985476 0.23728929 0.2607637 0.2492364 0.04262302
#> [12,] 0.31238752 0.4432714 0.2840189 0.23684282 0.2987134 0.2293411 0.04246271
#> [13,] 0.14583462 0.2023959 0.2160669 0.14156546 0.1711865 0.1623219 0.05880762
#> [14,] 0.24733918 0.4005275 0.2800513 0.59874650 0.6655884 0.5298783 0.10926696
#> [15,] 0.57344367 0.5049424 0.5355221 0.52497150 0.4252439 0.3180747 0.00932048
#> [16,] 0.30235380 0.2945294 0.2802306 0.36937254 0.4257341 0.3766998 0.03153780
#> [17,] 0.70514483 0.7137845 0.6486342 0.48108234 0.3323337 0.2722032 0.04453966
#> [18,] 0.42009663 0.3619669 0.2480768 0.09430693 0.0989298 0.1678797 0.02080476
#> [19,] 0.86434465 0.9591931 0.9551054 0.73012910 0.5834373 0.5423131 0.01566422
#> [20,] 0.20107905 0.4306339 0.5350665 0.43460636 0.2817383 0.1818971 0.04735414
#> [21,] 0.58009829 0.5776315 0.6991251 0.79596426 0.8975376 0.9527648 0.02710565
#> [22,] 0.74778928 0.8632167 0.9192558 0.84651477 0.7137975 0.5324582 0.02715962
#> [23,] 0.56683256 0.4346241 0.4493705 0.28830806 0.1512235 0.1644647 0.04380500
#> [24,] 0.16998005 0.2506619 0.4091437 0.33718355 0.1737580 0.1808004 0.03180849
#> [25,] 0.09864113 0.3951322 0.6809715 0.71566331 0.3805979 0.1505879 0.09473958
#>              V40       V41       V42          V43        V44         V45
#>  [1,] 0.24186808 0.1688835 0.1951167  0.218752330 0.10263548 0.046024223
#>  [2,] 0.06290157 0.1160951 0.1496649  0.138989017 0.13999262 0.068270607
#>  [3,] 0.45794168 0.3614981 0.2788900  0.300796906 0.29653071 0.413792299
#>  [4,] 0.42674549 0.1223597 0.3480063  0.448461299 0.47381133 0.366911163
#>  [5,] 0.22757963 0.2570673 0.2824247  0.228858310 0.11860343 0.121961571
#>  [6,] 0.33406786 0.3818596 0.4918700  0.372472989 0.17780380 0.179546812
#>  [7,] 0.38126507 0.2760092 0.1757567  0.294910890 0.26383326 0.272944748
#>  [8,] 0.31931706 0.5094254 0.5087544  0.433038196 0.38990925 0.435364715
#>  [9,] 0.44302627 0.3270167 0.4878316  0.459885216 0.47835517 0.522835807
#> [10,] 0.39129144 0.3535585 0.3645129  0.312547433 0.27366352 0.207697788
#> [11,] 0.24164979 0.2423230 0.2452852  0.233035807 0.19409266 0.150179783
#> [12,] 0.19398869 0.3920306 0.3548836 -0.007459756 0.12163331 0.194391633
#> [13,] 0.15881348 0.2293086 0.2491827  0.185985755 0.24298355 0.181502963
#> [14,] 0.47161699 0.4747627 0.2609444  0.214708138 0.21964502 0.202204749
#> [15,] 0.22299477 0.1456566 0.1368595  0.161124241 0.16055243 0.132894092
#> [16,] 0.29684404 0.2104317 0.1460521  0.101168855 0.05340139 0.092691113
#> [17,] 0.37614654 0.3674150 0.3376274  0.317425564 0.23139325 0.139762612
#> [18,] 0.17310990 0.2033649 0.2379412  0.230645693 0.15125009 0.073642383
#> [19,] 0.56326052 0.5185766 0.4730031  0.338500987 0.23026869 0.153731445
#> [20,] 0.30987352 0.3019074 0.2494219  0.180865521 0.18381092 0.176036021
#> [21,] 0.91352697 0.8560638 0.7681273  0.591132195 0.39404778 0.416717497
#> [22,] 0.33903747 0.2543329 0.1976744  0.226065479 0.15831145 0.176060192
#> [23,] 0.07587497 0.1093867 0.1696834  0.269101941 0.16333987 0.009957773
#> [24,] 0.13651308 0.2114908 0.2261130  0.179638475 0.11073970 0.094424912
#> [25,] 0.34486406 0.4323232 0.3692847  0.245391183 0.20135915 0.219577366
#>               V46        V47        V48         V49         V5         V50
#>  [1,] 0.007680476 0.01000314 0.03437259 0.011835912 0.12730704 0.003574137
#>  [2,] 0.023159226 0.07777381 0.07816860 0.023146676 0.12792878 0.010337871
#>  [3,] 0.349197912 0.21414126 0.13177512 0.087235101 0.03347802 0.029250093
#>  [4,] 0.106853899 0.11648912 0.17090129 0.104350252 0.12650443 0.027100673
#>  [5,] 0.130641074 0.14286553 0.11982966 0.064403610 0.07018156 0.014282566
#>  [6,] 0.131812655 0.08403705 0.08049904 0.044382210 0.27271381 0.015100441
#>  [7,] 0.231672365 0.15529971 0.09030072 0.062580987 0.04043431 0.024053903
#>  [8,] 0.421323326 0.27876198 0.18474270 0.105564426 0.09991865 0.035556916
#>  [9,] 0.274200043 0.17985112 0.20095953 0.107969072 0.09597524 0.031354220
#> [10,] 0.133851678 0.09935551 0.04377314 0.034753105 0.12389395 0.008123033
#> [11,] 0.142584871 0.14284295 0.09922808 0.058373162 0.10105058 0.023797352
#> [12,] 0.195698449 0.10105686 0.06271152 0.013190885 0.13178822 0.020492443
#> [13,] 0.121615762 0.08162729 0.06110301 0.049335702 0.09658569 0.008022769
#> [14,] 0.006118326 0.13389593 0.07515077 0.012198130 0.09828835 0.010195134
#> [15,] 0.083702356 0.06280686 0.07230181 0.038889204 0.01430376 0.012930503
#> [16,] 0.098847493 0.09361450 0.06251910 0.029630188 0.03858610 0.028696962
#> [17,] 0.100814745 0.07283475 0.05130065 0.024949188 0.07250272 0.016054845
#> [18,] 0.075762120 0.04932617 0.03418691 0.028669330 0.03029768 0.010484239
#> [19,] 0.087022860 0.09798186 0.08709286 0.054734725 0.05733077 0.014664241
#> [20,] 0.095086125 0.11874903 0.11296288 0.062370620 0.10597201 0.013220360
#> [21,] 0.451690997 0.35786696 0.23570420 0.148539976 0.07075420 0.055185905
#> [22,] 0.164013122 0.08943883 0.02655159 0.012935482 0.01475205 0.030660288
#> [23,] 0.014388421 0.03102911 0.03914298 0.007314145 0.02355685 0.020246856
#> [24,] 0.109028744 0.06792831 0.04679858 0.023789636 0.03754675 0.015662200
#> [25,] 0.179525671 0.18159895 0.16726646 0.114515651 0.06979592 0.029714601
#>               V51         V52         V53         V54         V55         V56
#>  [1,] 0.006029786 0.019785075 0.012403055 0.012477770 0.030854287 0.026216018
#>  [2,] 0.025848149 0.014041630 0.023957244 0.010343379 0.014343449 0.018008284
#>  [3,] 0.018309824 0.016519210 0.008367566 0.009334686 0.012107174 0.007819269
#>  [4,] 0.033267134 0.017208413 0.017976276 0.013235838 0.008072518 0.004390073
#>  [5,] 0.019464475 0.015184025 0.008491065 0.009621118 0.005124754 0.007703712
#>  [6,] 0.025664194 0.021669007 0.018860790 0.021155017 0.006389903 0.013107988
#>  [7,] 0.014990350 0.006555628 0.010547938 0.015046594 0.013245686 0.009512877
#>  [8,] 0.038516158 0.030369466 0.017523887 0.017452205 0.014764144 0.011412634
#>  [9,] 0.033043264 0.018677172 0.007942086 0.012141883 0.017688323 0.009182181
#> [10,] 0.005171578 0.011024343 0.015264063 0.013484713 0.007349728 0.013535441
#> [11,] 0.022047781 0.020339112 0.014629198 0.016981024 0.006496215 0.008622659
#> [12,] 0.016895586 0.020522946 0.022962802 0.034502991 0.018283583 0.015047077
#> [13,] 0.009272746 0.016480635 0.006939263 0.006838803 0.000959531 0.001454458
#> [14,] 0.002753628 0.023404398 0.016820121 0.008953628 0.017622650 0.025233316
#> [15,] 0.004218599 0.008173426 0.007926741 0.009809035 0.009355790 0.004572191
#> [16,] 0.016777262 0.014190057 0.021614259 0.021416736 0.008663295 0.009931083
#> [17,] 0.014198293 0.009170080 0.010061475 0.009655321 0.005972086 0.009471748
#> [18,] 0.011038838 0.012053092 0.009493458 0.003206338 0.005594326 0.002843218
#> [19,] 0.007619808 0.007044681 0.016065248 0.012128599 0.008429371 0.007764634
#> [20,] 0.018741925 0.012348501 0.008867128 0.010447059 0.006662964 0.006408367
#> [21,] 0.031293944 0.015406464 0.015019749 0.010502393 0.015164181 0.012558448
#> [22,] 0.008233528 0.010243821 0.003712232 0.010665728 0.009774077 0.002924572
#> [23,] 0.020095824 0.017505866 0.016553567 0.005590380 0.004500000 0.006772384
#> [24,] 0.008245111 0.011409218 0.008806249 0.009321557 0.011253821 0.008798822
#> [25,] 0.016692743 0.012728485 0.009336176 0.013108674 0.006807753 0.007013860
#>               V57         V58           V59         V6         V60         V7
#>  [1,] 0.006724706 0.007673303  0.0142520978 0.05677714 0.008976516 0.05166508
#>  [2,] 0.017834984 0.010857830  0.0128655504 0.11132138 0.013377939 0.10317463
#>  [3,] 0.007393327 0.007755163  0.0083951862 0.08599979 0.008386398 0.11776693
#>  [4,] 0.003915255 0.003098486  0.0056648907 0.20650752 0.005658777 0.21040179
#>  [5,] 0.006251212 0.007435965  0.0066698911 0.13209245 0.006889953 0.13014796
#>  [6,] 0.014485377 0.013564413  0.0137640579 0.17047244 0.007760512 0.07392201
#>  [7,] 0.007358862 0.011623075  0.0082736144 0.06022945 0.005695878 0.07030850
#>  [8,] 0.013589747 0.014005592  0.0138844345 0.11064895 0.012384674 0.12041146
#>  [9,] 0.010153121 0.009198174  0.0065309771 0.10431453 0.004177761 0.11538983
#> [10,] 0.010292665 0.012916926  0.0168537662 0.13791277 0.008923381 0.15572667
#> [11,] 0.010299521 0.010721712  0.0097123677 0.14132757 0.007936594 0.14874252
#> [12,] 0.014223547 0.012481361 -0.0002477906 0.20194983 0.004059862 0.16642122
#> [13,] 0.001472742 0.010532207  0.0060910111 0.15880919 0.002542185 0.17488010
#> [14,] 0.032511930 0.016565091  0.0095510995 0.23254000 0.007544502 0.23959772
#> [15,] 0.005304657 0.004134395  0.0073455207 0.07719182 0.005265634 0.05926197
#> [16,] 0.004393046 0.009363207  0.0019805015 0.02584121 0.003873777 0.03980415
#> [17,] 0.010216700 0.005304721  0.0061834316 0.11696616 0.007164245 0.11481991
#> [18,] 0.007355110 0.004057462  0.0034269463 0.07386982 0.004768134 0.07897132
#> [19,] 0.010254927 0.008571460  0.0078700901 0.09706516 0.005177886 0.07798715
#> [20,] 0.008456965 0.011413101  0.0168508134 0.12981609 0.012977059 0.18393536
#> [21,] 0.001549605 0.015579627  0.0126433714 0.09632327 0.013093504 0.18040514
#> [22,] 0.010595917 0.008203355  0.0072237615 0.05907199 0.009128664 0.12391273
#> [23,] 0.004125106 0.005172384  0.0197665422 0.06216103 0.010638081 0.04000577
#> [24,] 0.008825947 0.005991209  0.0056489840 0.06802597 0.002758120 0.10123838
#> [25,] 0.011598800 0.008662710  0.0067292030 0.08753228 0.004870821 0.15927216
#>               V8            V9
#>  [1,] 0.07108390  0.0300453324
#>  [2,] 0.09801517  0.2153403042
#>  [3,] 0.16358889  0.1959798859
#>  [4,] 0.24110499  0.3486309580
#>  [5,] 0.13609510  0.1912311912
#>  [6,] 0.18583161  0.0721500757
#>  [7,] 0.13387383  0.2558722642
#>  [8,] 0.14108224  0.1765118590
#>  [9,] 0.15107427  0.2556712353
#> [10,] 0.05036731  0.0340418053
#> [11,] 0.17763821  0.2502063459
#> [12,] 0.24097544  0.3910170690
#> [13,] 0.17397718  0.2044811700
#> [14,] 0.37388073  0.5600830129
#> [15,] 0.04348926  0.0666770842
#> [16,] 0.04471421  0.0901276496
#> [17,] 0.10678272  0.1004697441
#> [18,] 0.08279780  0.1086880048
#> [19,] 0.03363401 -0.0007199964
#> [20,] 0.17991663  0.1946482752
#> [21,] 0.19989483  0.0325601733
#> [22,] 0.11667783  0.0537496044
#> [23,] 0.11149617  0.1888831731
#> [24,] 0.11045364  0.1289733787
#> [25,] 0.21418822  0.2324539538
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768