Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.055644828 0.31991020 0.24869021 0.19867976 0.39328042 0.5076817
#>  [2,] 0.047742517 0.24450022 0.27986538 0.26094178 0.21270815 0.1076854
#>  [3,] 0.027200000 0.39970000 0.39410000 0.33090000 0.29260000 0.1760000
#>  [4,] 0.035408001 0.21799248 0.35609869 0.48657759 0.56098566 0.5929504
#>  [5,] 0.026445081 0.24796700 0.29376883 0.29635771 0.23177191 0.1114273
#>  [6,] 0.035384557 0.23304184 0.26410521 0.29757851 0.30964496 0.2753471
#>  [7,] 0.021270620 0.65125448 0.68461861 0.47682006 0.36310327 0.2738812
#>  [8,] 0.114775586 0.36034737 0.46211490 0.53975178 0.68005258 0.7801876
#>  [9,] 0.031443555 0.07001990 0.14513242 0.21927159 0.26662122 0.3232140
#> [10,] 0.042782585 0.32098471 0.39719291 0.45511830 0.44761211 0.3971328
#> [11,] 0.018974612 0.19624611 0.21938538 0.24059998 0.29538447 0.3656146
#> [12,] 0.016222017 0.24368060 0.28422331 0.33506865 0.30282130 0.2299216
#> [13,] 0.031242575 0.36877896 0.34138561 0.23249481 0.25569401 0.2796629
#> [14,] 0.010262151 0.03352509 0.10885392 0.14219996 0.22491499 0.3017789
#> [15,] 0.026500879 0.08790323 0.07793463 0.07638574 0.02714585 0.1024610
#> [16,] 0.019029842 0.13038952 0.08870005 0.03138103 0.13592842 0.3081068
#> [17,] 0.024168786 0.08710219 0.09608152 0.16862052 0.16943022 0.1511705
#> [18,] 0.013807268 0.07913254 0.11807679 0.13907559 0.17206081 0.2052638
#> [19,] 0.022161532 0.11923678 0.14952439 0.18419650 0.20019905 0.1809711
#> [20,] 0.057545934 0.08797969 0.22505486 0.39497804 0.36869668 0.1679366
#> [21,] 0.031957966 0.09811061 0.07161297 0.10270015 0.19405994 0.3450997
#> [22,] 0.025357513 0.24524207 0.28881634 0.34681962 0.55795717 0.6814011
#> [23,] 0.007909813 0.24049961 0.23365753 0.19133010 0.25416948 0.2571227
#> [24,] 0.018677262 0.10536470 0.10575332 0.11555475 0.13074574 0.1589923
#> [25,] 0.014202023 0.11196842 0.10489871 0.10618234 0.21487145 0.2783703
#>              V15        V16       V17       V18       V19            V2
#>  [1,] 0.22871368 0.37480735 0.5301068 0.5470620 0.7388254  0.0570400841
#>  [2,] 0.08066448 0.15099438 0.1762871 0.2359159 0.1997391  0.0361737974
#>  [3,] 0.17390000 0.20430000 0.2088000 0.2678000 0.2434000  0.0378000000
#>  [4,] 0.67420055 0.79295135 0.9319446 0.9554107 0.9083003  0.0564445543
#>  [5,] 0.16900121 0.20919540 0.2457274 0.2739476 0.2395007  0.0435554570
#>  [6,] 0.25254314 0.29714844 0.2873879 0.2257107 0.2879088  0.0629037976
#>  [7,] 0.18403384 0.09594454 0.1707669 0.3820724 0.6091304  0.0275379535
#>  [8,] 0.97332913 1.06479172 0.8853571 0.8823776 0.9135391  0.1244717779
#>  [9,] 0.41944012 0.51125426 0.6426776 0.7329367 0.8330378  0.0281049614
#> [10,] 0.31676490 0.28193903 0.2933866 0.3185855 0.4442282  0.0489594620
#> [11,] 0.42442158 0.53107339 0.6236788 0.7025604 0.7710035  0.0263164292
#> [12,] 0.16983614 0.21294224 0.2495180 0.3776867 0.6012786  0.0270920270
#> [13,] 0.22275495 0.18117193 0.2381556 0.3678407 0.5660537  0.0555458408
#> [14,] 0.41530442 0.47587092 0.5428268 0.6331018 0.7106406  0.0074062640
#> [15,] 0.07312244 0.17749018 0.1563304 0.2954675 0.3724067  0.0326542253
#> [16,] 0.54571395 0.77484202 0.8281848 0.8024738 0.8604000  0.0117916816
#> [17,] 0.17852641 0.18345592 0.2088345 0.2027951 0.2452332  0.0409632980
#> [18,] 0.15371479 0.20691275 0.2859718 0.3569296 0.4046278  0.0095924872
#> [19,] 0.13378852 0.19884534 0.2865071 0.3501086 0.3086943  0.0209337478
#> [20,] 0.25087836 0.38493164 0.5663281 0.6726457 0.6228039  0.0526263754
#> [21,] 0.48159858 0.64311541 0.7592916 0.8042987 0.8105066  0.0293778656
#> [22,] 0.78494459 0.71378774 0.3719493 0.2362215 0.3075586  0.0793557899
#> [23,] 0.38326031 0.57964843 0.7790435 0.7599940 0.5114661 -0.0004233716
#> [24,] 0.17984979 0.18889794 0.1617718 0.1655555 0.1258069  0.0174123220
#> [25,] 0.27345029 0.28680438 0.2795128 0.3712444 0.5040086  0.0179330932
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.8249378 0.5877407 0.6736778 0.8175540 0.6886365 0.7499645 0.8149816
#>  [2,] 0.1618994 0.3188191 0.5901664 0.7781300 0.8601125 0.9527933 0.9969424
#>  [3,] 0.1839000 0.2802000 0.6172000 0.8015000 0.8313000 0.8440000 0.8494000
#>  [4,] 0.8348279 0.7406555 0.6333414 0.5097270 0.4147261 0.3026170 0.1592832
#>  [5,] 0.2312694 0.4233386 0.6699666 0.7898231 0.8260714 0.8806672 0.9134150
#>  [6,] 0.3813656 0.4260256 0.4493017 0.5036427 0.5485922 0.5730515 0.6675234
#>  [7,] 0.7336729 0.8072461 0.8722017 0.8909530 0.9557021 0.8510875 0.7701319
#>  [8,] 1.0078431 0.8787783 0.7886489 0.9115403 0.8549131 0.5711270 0.5043922
#>  [9,] 0.8815600 0.8715630 0.8922179 0.8936196 0.8632470 0.7398276 0.5677450
#> [10,] 0.6597990 0.8063455 0.8263078 0.7576126 0.7269364 0.7153104 0.8184232
#> [11,] 0.8210720 0.8914547 0.9270263 0.9449247 0.9024982 0.8348527 0.8346590
#> [12,] 0.7087990 0.8078591 0.7995287 0.8322417 0.9452325 0.9479366 0.9579792
#> [13,] 0.8172999 0.7924035 0.4047954 0.1858614 0.2227646 0.3692557 0.6952734
#> [14,] 0.7931760 0.9526180 1.0015670 0.9704033 0.9589688 0.8848825 0.8099850
#> [15,] 0.3505672 0.2801215 0.1407181 0.1363274 0.2769133 0.1375583 0.2448712
#> [16,] 0.8692720 0.9299720 0.9336992 0.9388353 0.9731974 0.9439127 0.9459835
#> [17,] 0.2940295 0.2270855 0.2301008 0.2978983 0.4110544 0.4199764 0.3859891
#> [18,] 0.4618306 0.5630117 0.6449157 0.6552952 0.7476494 0.9460702 0.9493996
#> [19,] 0.4152233 0.5298531 0.6122456 0.7697252 0.8349906 0.8784104 0.8664040
#> [20,] 0.3907988 0.2613631 0.2311656 0.2543564 0.2427631 0.2871666 0.2895534
#> [21,] 0.8265722 0.9803857 0.9685433 0.9302284 0.9216480 0.8954786 0.9002809
#> [22,] 0.4575061 0.5452155 0.3696557 0.1612955 0.5337655 0.6273650 0.7373377
#> [23,] 0.4004608 0.5407060 0.5425220 0.5719342 0.7511101 0.7957186 0.7351069
#> [24,] 0.1400915 0.2327779 0.2494038 0.3876043 0.6180789 0.8937600 1.0073704
#> [25,] 0.6211123 0.6082111 0.5978055 0.5979225 0.5551582 0.7104468 0.8756442
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.7454195 0.6696193 0.4365128 0.066703988 0.6693135 0.7279438 0.8176003
#>  [2,] 0.9642777 1.0000000 0.8733485 0.020643131 0.7075989 0.7711176 0.7262889
#>  [3,] 0.9168000 1.0000000 0.7896000 0.048800000 0.5371000 0.6472000 0.6505000
#>  [4,] 0.1485161 0.1557585 0.1261893 0.061208415 0.1831012 0.2561859 0.3336395
#>  [5,] 0.9489962 0.9260269 0.7954135 0.037332966 0.6311296 0.6126535 0.5281194
#>  [6,] 0.7891788 0.8725294 0.8371196 0.064983798 0.8123389 0.7583084 0.6802501
#>  [7,] 0.7925660 0.7595066 0.5610593 0.048622545 0.3835287 0.3508687 0.3560595
#>  [8,] 0.5515954 0.5953935 0.5439805 0.074206374 0.3876020 0.3095691 0.2554457
#>  [9,] 0.4336535 0.3139552 0.3055616 0.040202138 0.3568658 0.3024186 0.3561280
#> [10,] 0.8779446 0.7876572 0.6243811 0.053351295 0.4800976 0.3806703 0.3349759
#> [11,] 0.8768372 0.8639598 0.7295094 0.028634087 0.5614087 0.3850945 0.3155621
#> [12,] 0.9792697 0.9800102 0.8715741 0.034537999 0.6334506 0.3927639 0.2900925
#> [13,] 0.7840470 0.7855045 0.7617544 0.056329817 0.7672409 0.4796171 0.4595878
#> [14,] 0.6678502 0.4549936 0.4763590 0.005299614 0.6268129 0.5031739 0.3174378
#> [15,] 0.4501722 0.7432235 0.9090775 0.059867733 0.7790739 0.5080179 0.2401215
#> [16,] 0.9543151 0.7742784 0.4894776 0.031068828 0.3638573 0.4013535 0.4163778
#> [17,] 0.4118559 0.5761479 0.6900217 0.032082268 0.7065149 0.7366396 0.6915736
#> [18,] 0.9892435 0.9289596 0.7804413 0.012276187 0.6972038 0.5819075 0.4383830
#> [19,] 0.8282263 0.7913078 0.7823678 0.025094189 0.7071349 0.6490033 0.6158981
#> [20,] 0.3045965 0.4331280 0.6969888 0.084348503 0.7524057 0.8521676 0.7968602
#> [21,] 0.7712847 0.5638516 0.3678004 0.045910508 0.4502381 0.4758375 0.3483400
#> [22,] 0.6725019 0.8048102 0.7942829 0.054483064 0.8323036 0.7838010 0.5048744
#> [23,] 0.7524542 0.8089285 0.6408134 0.006781190 0.4538671 0.4047601 0.3760508
#> [24,] 0.8281112 0.6314285 0.5360726 0.014696622 0.3881338 0.2957049 0.1313180
#> [25,] 0.9039392 0.7674409 0.5732078 0.026568982 0.4628563 0.4595116 0.3890272
#>              V33        V34        V35        V36        V37        V38
#>  [1,] 0.71141608 0.57749994 0.35086827 0.13585210 0.12419598 0.39346722
#>  [2,] 0.52209674 0.18632780 0.10599210 0.14126249 0.10868315 0.13920653
#>  [3,] 0.49590000 0.21750000 0.09900000 0.04340000 0.17080000 0.19790000
#>  [4,] 0.32910679 0.37235832 0.39246202 0.25925935 0.20680837 0.32317465
#>  [5,] 0.38755793 0.15893410 0.08808352 0.09378486 0.15550520 0.16014647
#>  [6,] 0.51827718 0.43975581 0.38894486 0.33915001 0.33863741 0.34967755
#>  [7,] 0.58541432 0.67178530 0.56821788 0.48644725 0.47147661 0.37274181
#>  [8,] 0.09930501 0.14257716 0.14188708 0.14033937 0.15270249 0.10314970
#>  [9,] 0.45528537 0.42221131 0.36433151 0.31170100 0.24256386 0.32477801
#> [10,] 0.29714191 0.24113159 0.09704849 0.18470584 0.24269916 0.24622268
#> [11,] 0.21554386 0.12439156 0.15306953 0.19031160 0.21367871 0.26887198
#> [12,] 0.21638931 0.18269136 0.17489211 0.09249443 0.07069309 0.09482160
#> [13,] 0.67907533 0.68825512 0.60278403 0.49767861 0.48342707 0.52110528
#> [14,] 0.37434118 0.33416726 0.09881638 0.05938255 0.02862578 0.03605532
#> [15,] 0.51881064 0.57896357 0.30624449 0.40125447 0.58570500 0.61116075
#> [16,] 0.34476077 0.28538632 0.24361679 0.14181167 0.24907364 0.20536817
#> [17,] 0.71466185 0.85371205 0.88150243 0.84151315 0.68666108 0.47134881
#> [18,] 0.28422753 0.09567282 0.02595906 0.20415335 0.20856013 0.28084715
#> [19,] 0.67899997 0.68405818 0.64528582 0.55629407 0.46314603 0.36149949
#> [20,] 0.63081193 0.65654856 0.72385701 0.81200129 0.86936034 0.98419888
#> [21,] 0.29548452 0.29306241 0.17302700 0.06589737 0.12979959 0.12100971
#> [22,] 0.14416925 0.27632936 0.50295546 0.46755530 0.16601788 0.25838161
#> [23,] 0.27348742 0.11697468 0.30489356 0.41502523 0.34467946 0.17334433
#> [24,] 0.14269021 0.17325510 0.23208761 0.19539782 0.05870843 0.09046810
#> [25,] 0.35643020 0.25418982 0.27897417 0.38628881 0.39804347 0.32915485
#>              V39            V4        V40        V41        V42        V43
#>  [1,] 0.59601283  0.0954488604 0.42542790 0.46861728 0.51382638 0.43515109
#>  [2,] 0.18929369  0.0352926402 0.03294799 0.15061276 0.04391011 0.06740844
#>  [3,] 0.18800000  0.0848000000 0.11080000 0.17020000 0.05850000 0.06380000
#>  [4,] 0.30310165  0.0806479477 0.13060768 0.09320397 0.15816984 0.15911743
#>  [5,] 0.17345243  0.0397616186 0.14451439 0.13429743 0.15170772 0.19781399
#>  [6,] 0.34186178  0.0666006708 0.33909328 0.48573083 0.52307234 0.44325031
#>  [7,] 0.38483966  0.0642461841 0.37475069 0.25016191 0.20912460 0.23410893
#>  [8,] 0.15233803  0.0006888166 0.23980490 0.16610665 0.13118504 0.11770782
#>  [9,] 0.30578415  0.0588747680 0.26976716 0.24754137 0.29840593 0.24728197
#> [10,] 0.30882300  0.0583395353 0.26854899 0.15443352 0.24733522 0.31488008
#> [11,] 0.27020733  0.0468608065 0.25921874 0.23921291 0.22892865 0.19777136
#> [12,] 0.09844833  0.0348217880 0.09564282 0.06050293 0.10087091 0.18730049
#> [13,] 0.45397128  0.0833883825 0.34930899 0.39216437 0.37849553 0.32131385
#> [14,] 0.16753971  0.0454249576 0.21981146 0.16444074 0.05932737 0.01196274
#> [15,] 0.56219573  0.0659172304 0.35900339 0.21882161 0.08347601 0.02790015
#> [16,] 0.11828242  0.0305130656 0.11293058 0.07444689 0.04613756 0.08102691
#> [17,] 0.35697665  0.0370087721 0.40630192 0.40205532 0.38444789 0.24050995
#> [18,] 0.27188566  0.0141289855 0.14792414 0.05865904 0.20035203 0.18946385
#> [19,] 0.27026387  0.0304030885 0.30154268 0.26044877 0.26489022 0.23383391
#> [20,] 0.90026013  0.0210557203 0.76187045 0.75168547 0.66307114 0.53952927
#> [21,] 0.08903467  0.0516114076 0.18978991 0.13564385 0.09280408 0.04319036
#> [22,] 0.29542534  0.0297614702 0.38896181 0.30425299 0.12465092 0.17320949
#> [23,] 0.11033116 -0.0041868975 0.16544276 0.18588536 0.17409643 0.13323731
#> [24,] 0.14349591  0.0213820388 0.21767016 0.29135425 0.26772499 0.18459324
#> [25,] 0.38276637  0.0330602087 0.27530484 0.23318199 0.25859001 0.20530936
#>               V44          V45          V46         V47         V48
#>  [1,] 0.380538472  0.502086072  0.502145073 0.292486743 0.226029382
#>  [2,] 0.139523774  0.091985574  0.060535598 0.099874558 0.106099853
#>  [3,] 0.139100000  0.063800000  0.058100000 0.064100000 0.104400000
#>  [4,] 0.060678455  0.086265635  0.060463534 0.050482917 0.052405987
#>  [5,] 0.127775010  0.103776292  0.118330497 0.163968205 0.143304963
#>  [6,] 0.371640388  0.469429455  0.398347051 0.262793511 0.177168052
#>  [7,] 0.208178942  0.153459931  0.096467909 0.097397098 0.118764723
#>  [8,] 0.233763289  0.131224313  0.056764379 0.081887013 0.093310907
#>  [9,] 0.185793209  0.157207670  0.091941533 0.071815845 0.059274387
#> [10,] 0.334223068  0.309191915  0.191477956 0.115815067 0.088557699
#> [11,] 0.195062740  0.125274230  0.107542410 0.111239450 0.081772072
#> [12,] 0.215872137  0.200781482  0.212044083 0.168128736 0.104407996
#> [13,] 0.217919018  0.188990939  0.154698272 0.094317231 0.099157955
#> [14,] 0.075665097  0.060126770  0.014099424 0.046917756 0.036837907
#> [15,] 0.007820003 -0.009015073 -0.006914563 0.009063007 0.004928327
#> [16,] 0.113208343  0.103216628  0.042630539 0.042884231 0.039952729
#> [17,] 0.228379059  0.185382308  0.152909278 0.129354947 0.082798352
#> [18,] 0.172398750  0.105994394  0.049792981 0.090356904 0.086859438
#> [19,] 0.157489917  0.111668594  0.102537654 0.065433235 0.055777007
#> [20,] 0.434161534  0.163273930  0.131855141 0.163260565 0.122794324
#> [21,] 0.128683891  0.119120692  0.064653797 0.104417685 0.075592194
#> [22,] 0.135330208  0.098581153  0.076012446 0.061690832 0.023289109
#> [23,] 0.108493353  0.083100880  0.085051904 0.061374117 0.051100092
#> [24,] 0.075981986  0.077994218  0.086218993 0.040965525 0.030130799
#> [25,] 0.124884021  0.086569831  0.084162614 0.067233455 0.036927907
#>                 V49          V5         V50          V51         V52
#>  [1,]  0.1272880314 0.206681789 0.072244595  0.021556627 0.018314671
#>  [2,]  0.0560087342 0.044261360 0.022245633  0.013170461 0.013782458
#>  [3,]  0.0732000000 0.112700000 0.027500000  0.014600000 0.009100000
#>  [4,]  0.0324794995 0.119376559 0.017728427  0.018892189 0.015137946
#>  [5,]  0.0835820482 0.047350958 0.026182053  0.018555131 0.014241904
#>  [6,]  0.1084187503 0.079584168 0.037574533  0.027179622 0.022352312
#>  [7,]  0.0514532839 0.043231213 0.021605140  0.011452204 0.013415947
#>  [8,]  0.0768408567 0.058475268 0.007916032  0.026271271 0.017888712
#>  [9,]  0.0330389716 0.099448754 0.011051236  0.010481713 0.011325252
#> [10,]  0.0565204919 0.068767731 0.016813128  0.017700572 0.011241585
#> [11,]  0.0491623194 0.101335227 0.018247277  0.016135456 0.013633630
#> [12,]  0.0526011956 0.068042792 0.014119580  0.020734282 0.014743139
#> [13,]  0.0496723272 0.136671706 0.019226276  0.018680683 0.014872333
#> [14,]  0.0249958333 0.075870739 0.009384773  0.016108560 0.002614139
#> [15,] -0.0003407681 0.051009811 0.001250447 -0.001014090 0.008150658
#> [16,]  0.0217877050 0.075276281 0.004342600  0.016885222 0.011010147
#> [17,]  0.0421626530 0.059316849 0.016069694  0.016545066 0.007414965
#> [18,]  0.0346774280 0.006236398 0.013402895  0.002509011 0.004743464
#> [19,]  0.0182963506 0.048095941 0.013014318  0.009597767 0.006773740
#> [20,]  0.0581062726 0.051734907 0.019264565  0.013305087 0.011527853
#> [21,]  0.0293912148 0.081559637 0.017466570  0.004038867 0.014099933
#> [22,]  0.0115022450 0.167607267 0.014812436  0.008167704 0.018145077
#> [23,]  0.0338230299 0.010280855 0.017168333  0.011370244 0.009829786
#> [24,]  0.0286836941 0.055222500 0.009659869  0.012186301 0.009078110
#> [25,]  0.0134906395 0.043406027 0.021684154  0.009012814 0.005722003
#>               V53         V54         V55         V56          V57         V58
#>  [1,] 0.027681315 0.019032055 0.035261857 0.030049529 2.816188e-02 0.034544040
#>  [2,] 0.004333562 0.006981551 0.006222399 0.008317615 1.162566e-02 0.011425662
#>  [3,] 0.004500000 0.004300000 0.004300000 0.009800000 5.400000e-03 0.005100000
#>  [4,] 0.013839130 0.017320429 0.017987238 0.009583398 6.863410e-03 0.009606714
#>  [5,] 0.009641916 0.018354696 0.004318125 0.009183944 6.860208e-03 0.011304222
#>  [6,] 0.013140303 0.011657044 0.010428445 0.008962720 7.618635e-03 0.009273371
#>  [7,] 0.008637278 0.019938401 0.012364480 0.010882194 6.516139e-03 0.016287525
#>  [8,] 0.014847780 0.015015941 0.021437434 0.003042623 1.194972e-03 0.010539188
#>  [9,] 0.007826705 0.010221719 0.007807893 0.011468245 6.275349e-03 0.007069835
#> [10,] 0.006549978 0.007745052 0.006418998 0.006102005 5.062993e-03 0.005775509
#> [11,] 0.013443151 0.014517948 0.007524558 0.009597752 6.608454e-03 0.010011792
#> [12,] 0.005821855 0.003342418 0.003958573 0.004954287 3.161646e-03 0.003801400
#> [13,] 0.014492827 0.015096803 0.006244588 0.008068176 5.668040e-03 0.006896152
#> [14,] 0.008061973 0.011505551 0.008405206 0.005791454 6.146287e-05 0.002112367
#> [15,] 0.024613054 0.018661170 0.008102905 0.010870884 1.317411e-02 0.006923212
#> [16,] 0.008260768 0.007953356 0.004026874 0.006988309 5.367673e-03 0.004908878
#> [17,] 0.012693008 0.010047550 0.011132651 0.007148401 8.750337e-03 0.008950303
#> [18,] 0.006712673 0.006174771 0.003183804 0.005926360 3.667843e-03 0.001724620
#> [19,] 0.006110571 0.006919151 0.006804761 0.007497207 8.096589e-03 0.007288031
#> [20,] 0.017118861 0.002260064 0.003086701 0.002937795 2.199989e-04 0.018399567
#> [21,] 0.010913579 0.003675945 0.003930825 0.016001611 1.276366e-02 0.001618534
#> [22,] 0.007894299 0.008002245 0.004707424 0.012417960 5.108458e-03 0.004698789
#> [23,] 0.005480485 0.010893873 0.008554294 0.007329534 7.342252e-03 0.003833890
#> [24,] 0.005692438 0.006852538 0.007510743 0.003384748 2.754211e-03 0.002135375
#> [25,] 0.009400109 0.008329967 0.006423044 0.006468729 8.162870e-03 0.009743626
#>               V59         V6         V60          V7         V8         V9
#>  [1,] 0.019955708 0.12568597 0.008558727 0.165329228 0.10942970 0.24277240
#>  [2,] 0.007364009 0.11232893 0.003906379 0.119719357 0.11421609 0.12069293
#>  [3,] 0.006500000 0.11030000 0.010300000 0.134900000 0.23370000 0.31130000
#>  [4,] 0.010521611 0.10025874 0.006808800 0.118330190 0.12550270 0.11833837
#>  [5,] 0.015657195 0.16666122 0.012222814 0.192353090 0.15572639 0.20569518
#>  [6,] 0.008700219 0.11452276 0.006662398 0.114516096 0.15341675 0.22892572
#>  [7,] 0.014710831 0.05156075 0.007430168 0.171659367 0.32930310 0.52198285
#>  [8,] 0.009993951 0.16351279 0.014177602 0.169152120 0.13795507 0.22542295
#>  [9,] 0.009134042 0.10537025 0.006198916 0.119442489 0.09608123 0.11496011
#> [10,] 0.005576757 0.11588905 0.005654206 0.126236977 0.16447405 0.25087144
#> [11,] 0.006252550 0.13499089 0.004662491 0.138249844 0.13111883 0.17457558
#> [12,] 0.002125395 0.08373535 0.003879931 0.118051806 0.13638317 0.19580578
#> [13,] 0.009022652 0.08219080 0.004942314 0.090472686 0.17129807 0.39197872
#> [14,] 0.002487120 0.08865717 0.002625530 0.099967307 0.04544869 0.04063589
#> [15,] 0.012474010 0.09202581 0.010397953 0.004628277 0.04763579 0.10227759
#> [16,] 0.006235418 0.12455576 0.004528664 0.116669142 0.07124571 0.12540161
#> [17,] 0.012275148 0.09091880 0.004855300 0.102020035 0.08191798 0.05999983
#> [18,] 0.003730687 0.03441803 0.003756467 0.087765095 0.07686761 0.07125887
#> [19,] 0.006182701 0.09313569 0.006753894 0.112165969 0.10848163 0.11289548
#> [20,] 0.014935237 0.08819908 0.009698818 0.081154811 0.10445443 0.16131067
#> [21,] 0.006453808 0.08361924 0.005115832 0.117117003 0.12416574 0.12790379
#> [22,] 0.008189464 0.37243975 0.005459758 0.364605377 0.35950138 0.35676058
#> [23,] 0.005305711 0.04972476 0.003116875 0.068803457 0.11162120 0.16006189
#> [24,] 0.003649178 0.07191911 0.002004716 0.074298521 0.06222761 0.10495636
#> [25,] 0.004140462 0.04876095 0.004312304 0.053950671 0.04030642 0.08394934
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696