Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13       V14
#>  [1,] 0.027844663 0.28515682 0.32056611 0.34100566 0.30754365 0.2884782
#>  [2,] 0.030811438 0.17235645 0.20379425 0.23681425 0.31933568 0.3935238
#>  [3,] 0.010770411 0.24327906 0.34733129 0.31586527 0.29229032 0.2218180
#>  [4,] 0.033278427 0.09454232 0.19570683 0.30257910 0.38978114 0.4549329
#>  [5,] 0.115409655 0.16803556 0.26797491 0.21809077 0.27880043 0.3306475
#>  [6,] 0.021729315 0.36778647 0.34933438 0.30570435 0.43375679 0.6305275
#>  [7,] 0.044465783 0.24326681 0.14073712 0.12989803 0.22062871 0.1499999
#>  [8,] 0.058524648 0.13052145 0.22214623 0.22845722 0.33413129 0.3359196
#>  [9,] 0.101191106 0.61059092 0.51829244 0.23412309 0.21040672 0.3111457
#> [10,] 0.030210120 0.36956884 0.44387043 0.44517001 0.41614371 0.3148765
#> [11,] 0.018651773 0.15065518 0.07929598 0.01294939 0.10412301 0.2462339
#> [12,] 0.029320910 0.25651831 0.30834994 0.31196975 0.26041388 0.1334432
#> [13,] 0.017660296 0.34822052 0.30198023 0.29553321 0.25442714 0.1805038
#> [14,] 0.025847720 0.08809972 0.07141790 0.12206890 0.15039090 0.1163882
#> [15,] 0.007743813 0.08197112 0.08804003 0.11337918 0.16221225 0.2052705
#> [16,] 0.018594216 0.13066446 0.13471379 0.08328086 0.09814058 0.1151586
#> [17,] 0.013784862 0.17045067 0.12373762 0.11387965 0.23578379 0.3275222
#> [18,] 0.013023685 0.12597587 0.15092320 0.19861431 0.19541451 0.1688816
#> [19,] 0.019931092 0.05259654 0.06781194 0.11531633 0.20849045 0.3873321
#> [20,] 0.025986649 0.11174138 0.09312448 0.07591871 0.04628588 0.1152791
#> [21,] 0.019176585 0.13772709 0.16877385 0.23821631 0.33071751 0.3397443
#> [22,] 0.014578894 0.04771959 0.11021257 0.13035346 0.21433821 0.2694974
#> [23,] 0.031525614 0.12148970 0.11156763 0.13334367 0.21648298 0.3586816
#> [24,] 0.017462620 0.10746147 0.13228461 0.11437293 0.13374791 0.1228465
#> [25,] 0.011874920 0.03375920 0.09272599 0.12840717 0.22267330 0.3170433
#>              V15        V16        V17        V18       V19          V2
#>  [1,] 0.24893912 0.25954722 0.24961809 0.22385447 0.3338386 0.028775708
#>  [2,] 0.39361670 0.48786338 0.57619155 0.65415380 0.7844593 0.027842159
#>  [3,] 0.13661176 0.09591851 0.23010437 0.33017491 0.4244790 0.030400194
#>  [4,] 0.58192431 0.57237106 0.73542508 0.85501986 0.9368289 0.047603814
#>  [5,] 0.37477242 0.41350012 0.41802555 0.43789570 0.5195082 0.212673857
#>  [6,] 0.62754136 0.70561999 0.79389035 0.82963105 0.8736125 0.076190999
#>  [7,] 0.06299525 0.10343802 0.07981309 0.04465015 0.1855990 0.026652400
#>  [8,] 0.31648880 0.34388190 0.38177515 0.21222456 0.2828404 0.089081424
#>  [9,] 0.59655708 0.59672871 0.20654270 0.16098949 0.4484467 0.099867099
#> [10,] 0.22209407 0.23144979 0.27040559 0.34944430 0.5659955 0.044765222
#> [11,] 0.40842251 0.59929056 0.71730337 0.80025778 0.9074058 0.018432513
#> [12,] 0.13669819 0.17689252 0.20565978 0.25705089 0.2538048 0.023858090
#> [13,] 0.12208205 0.22225367 0.24534079 0.28102551 0.2024691 0.021144698
#> [14,] 0.15297360 0.14542093 0.17172974 0.20797427 0.2604422 0.051386968
#> [15,] 0.17805541 0.17990442 0.16598115 0.17686747 0.1738843 0.011885046
#> [16,] 0.10522374 0.14585484 0.22135833 0.24500956 0.2793280 0.011498254
#> [17,] 0.29743330 0.40493795 0.41940289 0.43664153 0.5520248 0.047440896
#> [18,] 0.15372037 0.23945932 0.38278919 0.45411856 0.3894311 0.009494023
#> [19,] 0.51141662 0.67750502 0.75773650 0.77774030 0.7823327 0.028398540
#> [20,] 0.08471607 0.18181969 0.15758122 0.29622767 0.3636257 0.032810990
#> [21,] 0.36758267 0.37471568 0.35912215 0.30534569 0.4125120 0.022653021
#> [22,] 0.29945663 0.26409870 0.26108521 0.33657524 0.4621270 0.006796289
#> [23,] 0.52145282 0.68738528 0.80544343 0.83861998 0.8395771 0.028442694
#> [24,] 0.09249902 0.13214151 0.17368729 0.23479594 0.3972446 0.014403738
#> [25,] 0.42870018 0.49091463 0.57663763 0.68002018 0.7528891 0.013091473
#>             V20        V21       V22       V23       V24       V25       V26
#>  [1,] 0.4421596 0.53840565 0.5313026 0.5418014 0.5719290 0.5864281 0.7222935
#>  [2,] 0.8487754 0.84064749 0.9014974 0.9518120 0.9126155 0.8917912 0.9193165
#>  [3,] 0.4651252 0.45608508 0.4258375 0.3626622 0.4474095 0.6145352 0.8300987
#>  [4,] 0.9441954 0.96390317 0.8426538 0.7621865 0.7569279 0.5173108 0.3797456
#>  [5,] 0.3460397 0.21091334 0.1619265 0.3889001 0.6376143 0.4993453 0.3077407
#>  [6,] 0.8864276 0.86359099 0.9101484 1.0381384 1.0260978 0.9176461 0.9392354
#>  [7,] 0.3058806 0.28177299 0.3356875 0.3615043 0.3087946 0.3442501 0.3439634
#>  [8,] 0.3861526 0.50603354 0.6143652 0.6098505 0.5875712 0.5376838 0.5458131
#>  [9,] 0.8813238 0.85766657 0.3801304 0.1504748 0.1277792 0.2196630 0.3352903
#> [10,] 0.7193109 0.80797636 0.8290039 0.7995031 0.8569483 0.8629397 0.8936440
#> [11,] 0.9566670 0.93040467 0.9298083 0.8470669 0.5662886 0.3393276 0.3226649
#> [12,] 0.2888492 0.40038754 0.6323783 0.7903620 0.8471363 0.9258126 0.9542954
#> [13,] 0.1539944 0.07224439 0.1337682 0.3194234 0.4849022 0.6445337 0.8007170
#> [14,] 0.2487731 0.22311244 0.1965875 0.2385838 0.2886809 0.3152256 0.3019496
#> [15,] 0.1861298 0.24624880 0.3717528 0.4738384 0.5995015 0.6686255 0.7298707
#> [16,] 0.4105175 0.55373553 0.7256621 0.8888440 0.8858372 0.8710510 0.9332059
#> [17,] 0.6345686 0.61677900 0.5388969 0.5300740 0.4714149 0.5795119 0.7276054
#> [18,] 0.3970410 0.46927711 0.6473708 0.7186187 0.7576791 0.8106531 0.8491957
#> [19,] 0.8497646 0.95576517 0.8785206 0.8268174 0.8794806 0.9043662 0.9491560
#> [20,] 0.3732760 0.30847936 0.1595670 0.1456219 0.2672748 0.1286877 0.2440048
#> [21,] 0.5111319 0.56515726 0.5692969 0.5720596 0.7234211 0.8381936 0.8634833
#> [22,] 0.5388125 0.51143251 0.5515470 0.6198980 0.6771889 0.8399342 0.9184892
#> [23,] 0.8372187 0.94075276 0.9425538 0.8812493 0.8705781 0.8695461 0.8810544
#> [24,] 0.3800362 0.28450551 0.2983095 0.3967106 0.5380421 0.5343251 0.5389425
#> [25,] 0.8030319 0.93045723 0.9734571 0.9492881 0.9103759 0.8246935 0.7613261
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.8665082 0.8807918 0.8654519 0.037424665 0.7901682 0.6377856 0.5961671
#>  [2,] 0.8857156 0.8021836 0.6406898 0.032393797 0.5431842 0.4267317 0.3677228
#>  [3,] 1.0044240 1.0481515 0.9007405 0.031799152 0.6613512 0.5577713 0.5290356
#>  [4,] 0.3607248 0.2998730 0.2488101 0.069879459 0.2209126 0.1494579 0.2323172
#>  [5,] 0.4142371 0.5451290 0.4941075 0.281471029 0.4146374 0.3624567 0.2402145
#>  [6,] 0.8583554 0.6256606 0.5603560 0.078448927 0.4392310 0.2668755 0.1737090
#>  [7,] 0.3946582 0.6980266 0.9775493 0.062446210 1.0508341 0.8812583 0.6345887
#>  [8,] 0.6097890 0.7842567 0.8333907 0.092364507 0.8147100 0.8856855 0.8384847
#>  [9,] 0.2342439 0.5917951 0.6045287 0.033569023 0.7392979 0.4097687 0.4260434
#> [10,] 0.9536821 0.9156399 0.7408615 0.039641907 0.5175636 0.3921707 0.3726444
#> [11,] 0.3272353 0.4783461 0.6493478 0.007138814 0.6448577 0.4290099 0.3589519
#> [12,] 0.9539126 0.9547939 0.8393233 0.032373344 0.7078791 0.7033457 0.5951610
#> [13,] 0.9270559 0.9533925 0.9296243 0.015338235 0.9889562 0.9659025 0.8597727
#> [14,] 0.2779389 0.4575808 0.6288008 0.035654340 0.7586904 0.7674290 0.6917422
#> [15,] 0.7474070 0.8407521 0.9329944 0.019153114 0.9005938 0.7492860 0.6524028
#> [16,] 0.9381746 0.8909722 0.8660291 0.017365275 0.7474166 0.6390865 0.5801223
#> [17,] 0.6702095 0.4314587 0.2673324 0.063460895 0.2719160 0.3244092 0.4507928
#> [18,] 0.7719205 0.7278751 0.7977091 0.029887003 0.8003580 0.7078963 0.6644769
#> [19,] 0.9145747 0.7184172 0.3954959 0.005363693 0.3003793 0.2966805 0.2460069
#> [20,] 0.4359882 0.7198759 0.8734288 0.056014124 0.7725958 0.4978557 0.2406864
#> [21,] 0.8742067 0.8059020 0.6953788 0.025628950 0.5291934 0.4099400 0.2911723
#> [22,] 0.9627270 0.9390613 0.7920997 0.015340598 0.6937472 0.6186909 0.4238999
#> [23,] 0.7657832 0.5810212 0.4427482 0.045352553 0.4786220 0.4765539 0.3587635
#> [24,] 0.5479989 0.7560676 0.9848699 0.016953974 0.9766720 0.9224349 0.8513094
#> [25,] 0.6301355 0.4708881 0.5237050 0.014523495 0.6180264 0.4437383 0.2558175
#>             V33       V34       V35        V36        V37        V38        V39
#>  [1,] 0.5169100 0.3590402 0.1489553 0.18555437 0.25373361 0.25139981 0.26305704
#>  [2,] 0.2650354 0.1855216 0.1745153 0.16903966 0.16390874 0.24411029 0.30076576
#>  [3,] 0.4716928 0.3557350 0.2546299 0.18832331 0.26114545 0.18855379 0.17314638
#>  [4,] 0.3728119 0.3128052 0.3562377 0.20328931 0.22231571 0.43026151 0.31651548
#>  [5,] 0.1394158 0.1945207 0.2846464 0.40997951 0.31502924 0.31512975 0.37841405
#>  [6,] 0.0968860 0.2569013 0.3042468 0.21127826 0.16695239 0.21798822 0.24573888
#>  [7,] 0.5197901 0.3899333 0.2582631 0.15207115 0.29279926 0.43365852 0.57190879
#>  [8,] 0.6438412 0.5868133 0.6254068 0.35518298 0.18307188 0.26988278 0.22250824
#>  [9,] 0.7069946 0.8338659 0.9378187 0.86477728 0.87372380 0.90072777 0.74171443
#> [10,] 0.3524113 0.2945791 0.1752965 0.21637324 0.22356214 0.23959437 0.29739933
#> [11,] 0.4443558 0.3768795 0.2925873 0.19383223 0.09856677 0.15832624 0.14559254
#> [12,] 0.4326814 0.2056455 0.1064932 0.09417899 0.14381374 0.15772668 0.17004977
#> [13,] 0.7165567 0.6331508 0.6602682 0.72508979 0.74625475 0.62908130 0.53039610
#> [14,] 0.7305783 0.8499193 0.9302351 0.90752892 0.70515692 0.55674939 0.50312141
#> [15,] 0.6749303 0.5850559 0.5161759 0.44122724 0.33443748 0.30382989 0.29829843
#> [16,] 0.5542285 0.5974769 0.6118604 0.54897326 0.45189746 0.33319610 0.21333825
#> [17,] 0.4365344 0.2948595 0.2506894 0.40203629 0.58429863 0.56863050 0.61136988
#> [18,] 0.7833601 0.8116884 0.7998243 0.73513298 0.58972346 0.44252292 0.28543793
#> [19,] 0.2139075 0.1680432 0.1434932 0.08517523 0.13524873 0.14322222 0.10816433
#> [20,] 0.5173878 0.5769543 0.3264907 0.41935864 0.58271831 0.62212525 0.58030319
#> [21,] 0.1930831 0.1711093 0.2229669 0.23177810 0.23508580 0.28781342 0.26467049
#> [22,] 0.3180182 0.1460613 0.1048132 0.30997951 0.30872186 0.28297248 0.28550230
#> [23,] 0.3245108 0.3183726 0.1985967 0.08075134 0.15720510 0.14247010 0.10191855
#> [24,] 0.7560131 0.6446404 0.5506339 0.48894387 0.48432637 0.24161231 0.09460323
#> [25,] 0.3051787 0.3013172 0.1062609 0.05623737 0.01977879 0.05090507 0.17944360
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.04771380 0.33331244 0.43041914 0.47284588 0.44438195 0.33781444
#>  [2,] 0.06653348 0.26536802 0.27477553 0.29430667 0.25230417 0.21805637
#>  [3,] 0.05119204 0.25543679 0.32837500 0.42110633 0.33373429 0.13078702
#>  [4,] 0.11821467 0.22572017 0.19146107 0.24630911 0.25971485 0.13080454
#>  [5,] 0.39489824 0.42403399 0.36132385 0.27890295 0.18048841 0.20409481
#>  [6,] 0.05422929 0.22356526 0.27304968 0.22150239 0.08890541 0.16730917
#>  [7,] 0.05605556 0.58752766 0.46411197 0.30886536 0.23161307 0.29690396
#>  [8,] 0.08929816 0.18601897 0.60192724 0.72852392 0.48034600 0.44182132
#>  [9,] 0.10707839 0.51018922 0.28248679 0.28816046 0.25965302 0.32137614
#> [10,] 0.03706316 0.26387738 0.11827331 0.19887397 0.28970617 0.33101380
#> [11,] 0.04602098 0.10861691 0.15862474 0.17968951 0.14163863 0.19794855
#> [12,] 0.04999270 0.08634057 0.13584231 0.11502868 0.14463898 0.15430246
#> [13,] 0.05006689 0.53540601 0.46659459 0.38141207 0.42487675 0.53260550
#> [14,] 0.01877546 0.47613444 0.46231692 0.45281180 0.29694735 0.26232699
#> [15,] 0.02682339 0.24432260 0.18304601 0.12405489 0.15770379 0.14340855
#> [16,] 0.03387484 0.20055881 0.17744823 0.10995365 0.12576276 0.08557595
#> [17,] 0.05581803 0.56262452 0.44060892 0.33814474 0.20807759 0.15195532
#> [18,] 0.03071766 0.37460270 0.34888265 0.30556608 0.27119000 0.19665878
#> [19,] 0.03232212 0.11067189 0.09521389 0.04005185 0.07191418 0.11342408
#> [20,] 0.06199347 0.38476500 0.23446439 0.11100965 0.05060610 0.02541626
#> [21,] 0.02101722 0.17189403 0.13597863 0.15678877 0.17086751 0.14850018
#> [22,] 0.01432118 0.13682432 0.11863753 0.26199139 0.24032625 0.15503479
#> [23,] 0.04146546 0.18294621 0.13930407 0.11188417 0.06858854 0.12311475
#> [24,] 0.03219581 0.13333204 0.14412645 0.17143574 0.26211378 0.36017691
#> [25,] 0.05980699 0.20302166 0.18379838 0.07571530 0.01662549 0.07551124
#>              V45        V46        V47        V48         V49         V5
#>  [1,] 0.37198889 0.26283368 0.20448378 0.18628986 0.113379638 0.06878085
#>  [2,] 0.19104398 0.17889165 0.13758267 0.09880815 0.057593188 0.13900784
#>  [3,] 0.15255721 0.14300687 0.16880770 0.14711300 0.064093853 0.08227185
#>  [4,] 0.07757606 0.02928627 0.03656821 0.05129957 0.031503132 0.11999205
#>  [5,] 0.28525988 0.23004984 0.18541581 0.15342838 0.069754943 0.36735547
#>  [6,] 0.16084432 0.14788582 0.10093981 0.07736719 0.040787662 0.11719271
#>  [7,] 0.42096860 0.42272330 0.35315198 0.25027432 0.126653452 0.04634371
#>  [8,] 0.61025308 0.59569627 0.33715517 0.15975058 0.071728362 0.09270460
#>  [9,] 0.45548966 0.33783047 0.13278079 0.07960484 0.043713167 0.09338392
#> [10,] 0.28740399 0.17117113 0.10902516 0.09528269 0.051861670 0.05217042
#> [11,] 0.18315488 0.07719838 0.07543431 0.07437448 0.051193748 0.06539352
#> [12,] 0.09431380 0.05882406 0.09336471 0.10309762 0.059747405 0.06065164
#> [13,] 0.49324084 0.31627781 0.20074484 0.19714499 0.110348207 0.07702290
#> [14,] 0.20626075 0.16720931 0.16158122 0.12098162 0.065479266 0.03858318
#> [15,] 0.10748800 0.06302154 0.03976075 0.03985063 0.019413619 0.02732326
#> [16,] 0.04054749 0.04411735 0.02886235 0.02505115 0.008393658 0.05617959
#> [17,] 0.10798377 0.07910158 0.06083933 0.04248080 0.035245745 0.06872993
#> [18,] 0.14562635 0.12836258 0.07843047 0.04763069 0.013050692 0.05120504
#> [19,] 0.07109971 0.04906194 0.07097110 0.05072057 0.041517985 0.04273138
#> [20,] 0.01517051 0.01058539 0.02802016 0.02705785 0.008972111 0.04926968
#> [21,] 0.09907140 0.05842609 0.04386933 0.03027527 0.010308020 0.06317504
#> [22,] 0.14305432 0.11288549 0.08103720 0.06865334 0.023708566 0.02990412
#> [23,] 0.12979339 0.09421374 0.11108766 0.07651376 0.032855835 0.08080633
#> [24,] 0.33941478 0.22240144 0.10612103 0.04712869 0.029481517 0.03451590
#> [25,] 0.06974981 0.01216893 0.04402631 0.03541426 0.024192616 0.08535428
#>               V50         V51         V52         V53          V54         V55
#>  [1,] 0.026399865 0.021432105 0.014648125 0.005817960 0.0073855013 0.007365399
#>  [2,] 0.029079532 0.018278614 0.016473476 0.017065994 0.0175149103 0.012255804
#>  [3,] 0.013352769 0.019456953 0.009940243 0.009746759 0.0098275480 0.001313728
#>  [4,] 0.016332995 0.018030104 0.027493314 0.016903871 0.0149830885 0.022063158
#>  [5,] 0.019809879 0.014212245 0.035850580 0.021808452 0.0159276921 0.020630619
#>  [6,] 0.021073135 0.023031391 0.015591017 0.019829070 0.0245630792 0.020174214
#>  [7,] 0.031385139 0.016806837 0.022797609 0.002892713 0.0005870897 0.005567987
#>  [8,] 0.035932690 0.046905509 0.041110083 0.024816769 0.0201111058 0.017183770
#>  [9,] 0.026788715 0.017446983 0.023161328 0.018687709 0.0236031093 0.023874062
#> [10,] 0.012954597 0.019396961 0.012250684 0.007191864 0.0093855259 0.006487924
#> [11,] 0.015523299 0.014649741 0.012704899 0.003841349 0.0035659864 0.003542891
#> [12,] 0.019729350 0.009867020 0.008726952 0.008217259 0.0105002975 0.005227866
#> [13,] 0.030509932 0.020158344 0.019101533 0.011071115 0.0088674316 0.004985197
#> [14,] 0.019855013 0.016622751 0.006814886 0.011592170 0.0091473313 0.010439956
#> [15,] 0.010388747 0.008973497 0.009201452 0.011378707 0.0063149006 0.006709625
#> [16,] 0.007939770 0.004019173 0.004839206 0.006133434 0.0056929838 0.005755754
#> [17,] 0.019606126 0.018207591 0.016433226 0.009393000 0.0177321249 0.016340762
#> [18,] 0.016899176 0.013747474 0.012725059 0.010682066 0.0106394292 0.007139034
#> [19,] 0.025975196 0.014620760 0.009846735 0.007777876 0.0046675920 0.005225975
#> [20,] 0.004019053 0.002151042 0.010326633 0.024209457 0.0187301118 0.010179484
#> [21,] 0.007500178 0.005146783 0.008386799 0.009982963 0.0078693707 0.009864433
#> [22,] 0.024847594 0.008257352 0.003095746 0.003576055 0.0039014770 0.002773133
#> [23,] 0.017472832 0.007306900 0.011780875 0.009451624 0.0051339837 0.005230067
#> [24,] 0.018190655 0.012041590 0.004404436 0.005401171 0.0110831792 0.002524297
#> [25,] 0.009101292 0.015532479 0.002948110 0.006817769 0.0112265589 0.009537264
#>               V56          V57         V58          V59         V6         V60
#>  [1,] 0.006950878 0.0044701601 0.006598446 0.0067302074 0.08565640 0.005248860
#>  [2,] 0.014157773 0.0109328637 0.014565224 0.0110001151 0.12693263 0.006170904
#>  [3,] 0.002595103 0.0013325952 0.006715163 0.0054600257 0.08202854 0.001292537
#>  [4,] 0.023396011 0.0053216068 0.016694937 0.0227939286 0.03932018 0.015805483
#>  [5,] 0.002707755 0.0134055811 0.015166606 0.0186257896 0.17178788 0.021767022
#>  [6,] 0.012033038 0.0119681470 0.015579053 0.0043380594 0.17427959 0.003062687
#>  [7,] 0.013189607 0.0117252891 0.003984850 0.0079652617 0.03201521 0.003154552
#>  [8,] 0.011745756 0.0108816330 0.012834382 0.0157710741 0.14914582 0.011568794
#>  [9,] 0.009615758 0.0099486708 0.011692847 0.0076054752 0.09558872 0.013640096
#> [10,] 0.005435868 0.0047508604 0.005357102 0.0056366402 0.11022863 0.005075318
#> [11,] 0.003694556 0.0021196025 0.005638500 0.0056944526 0.14159439 0.001123362
#> [12,] 0.006770875 0.0044888656 0.005205468 0.0069651899 0.09955260 0.007943567
#> [13,] 0.004842827 0.0070070433 0.002958862 0.0008381355 0.10177036 0.001075540
#> [14,] 0.008629331 0.0070154933 0.013727484 0.0110784029 0.07063656 0.003455963
#> [15,] 0.006921116 0.0065136597 0.002904272 0.0070345827 0.04499841 0.004847373
#> [16,] 0.005279603 0.0062071344 0.005536889 0.0039441719 0.08811947 0.003510691
#> [17,] 0.006392818 0.0082793488 0.012663496 0.0070051342 0.05696656 0.003706320
#> [18,] 0.007578595 0.0107995243 0.006249858 0.0054426010 0.08206158 0.009802801
#> [19,] 0.007329873 0.0061194930 0.004267544 0.0046116958 0.08607792 0.003233772
#> [20,] 0.010982894 0.0122677344 0.007165995 0.0124848405 0.08874076 0.010364117
#> [21,] 0.007507475 0.0055000080 0.006026911 0.0045513824 0.08759202 0.003117357
#> [22,] 0.005966507 0.0125279356 0.012239498 0.0036402108 0.05673403 0.005613749
#> [23,] 0.013289745 0.0102568375 0.002937758 0.0059163380 0.10090489 0.005599512
#> [24,] 0.002301869 0.0050614499 0.004072664 0.0067764017 0.04550300 0.001938550
#> [25,] 0.005673708 0.0000962714 0.002104270 0.0026887980 0.10305736 0.003185063
#>               V7         V8         V9
#>  [1,] 0.08395494 0.10850293 0.21667458
#>  [2,] 0.14905725 0.12394033 0.15001503
#>  [3,] 0.08909741 0.09286328 0.19986859
#>  [4,] 0.09252925 0.01826526 0.03769090
#>  [5,] 0.17871204 0.03406055 0.16561881
#>  [6,] 0.21395973 0.26526128 0.33908363
#>  [7,] 0.04752265 0.10159163 0.13312576
#>  [8,] 0.10169540 0.13474853 0.20660732
#>  [9,] 0.25854537 0.40961812 0.56677035
#> [10,] 0.13711245 0.17793467 0.27575848
#> [11,] 0.15242477 0.11945619 0.15715227
#> [12,] 0.11896964 0.13985973 0.18062026
#> [13,] 0.13116015 0.15978377 0.31570200
#> [14,] 0.08881178 0.07485020 0.06245922
#> [15,] 0.04543986 0.04519709 0.07015425
#> [16,] 0.08912337 0.09241655 0.13031297
#> [17,] 0.05635527 0.06305123 0.10250310
#> [18,] 0.12080615 0.09488146 0.06578629
#> [19,] 0.15473461 0.11725541 0.10313109
#> [20,] 0.02061349 0.06929631 0.13234353
#> [21,] 0.10843191 0.09496010 0.10182582
#> [22,] 0.06566383 0.02199883 0.06057060
#> [23,] 0.11407480 0.10786117 0.13792919
#> [24,] 0.09253621 0.08372596 0.08631752
#> [25,] 0.11092766 0.05482373 0.03924164
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841