Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.014829788  0.20457553 0.25023340 0.29719062 0.28002378 0.24566705
#>  [2,] 0.055923187 -0.02827393 0.13379757 0.25063846 0.31215163 0.37007800
#>  [3,] 0.050110160  0.31189166 0.25119519 0.20638206 0.38580997 0.47620858
#>  [4,] 0.076731039  0.51435905 0.36576764 0.15500987 0.23392785 0.32550152
#>  [5,] 0.022515322  0.30437121 0.26135695 0.27639883 0.23817302 0.19838786
#>  [6,] 0.035744849  0.26999790 0.24470177 0.27443753 0.32933727 0.28385968
#>  [7,] 0.086660725  0.63910035 0.49328478 0.25261872 0.14888163 0.39760461
#>  [8,] 0.029242639  0.33589471 0.39779997 0.42144778 0.39966658 0.33638551
#>  [9,] 0.033024989  0.21914755 0.30687378 0.29346533 0.25011197 0.32298746
#> [10,] 0.024054578  0.19066985 0.15708772 0.11894044 0.13307939 0.18221505
#> [11,] 0.041310139  0.21780848 0.34719460 0.45842593 0.51628946 0.61046163
#> [12,] 0.032419244  0.29771674 0.33728930 0.35951525 0.27915530 0.17452323
#> [13,] 0.017007537  0.49331612 0.47992085 0.43932233 0.54697289 0.72119422
#> [14,] 0.034429287  0.24864003 0.31657776 0.33876731 0.29721844 0.15106606
#> [15,] 0.011908649  0.10198976 0.10925885 0.14730505 0.17259860 0.23532485
#> [16,] 0.010892967  0.08428406 0.02932889 0.12480636 0.12567689 0.09235337
#> [17,] 0.023442835  0.11533536 0.12897661 0.06580575 0.10127152 0.10240163
#> [18,] 0.015826507  0.13816959 0.11166220 0.09582386 0.07152671 0.07184257
#> [19,] 0.005843194  0.07937686 0.14491407 0.15567389 0.17390209 0.22311080
#> [20,] 0.033235199  0.32555578 0.24952201 0.11606872 0.18694244 0.47987342
#> [21,] 0.013120257  0.09569063 0.06856410 0.10440422 0.23150321 0.25496005
#> [22,] 0.025760688  0.06974825 0.09403144 0.22510938 0.16963720 0.10824877
#> [23,] 0.019500000  0.11910000 0.15220000 0.13220000 0.14340000 0.12440000
#> [24,] 0.022300000  0.14870000 0.11560000 0.16540000 0.38330000 0.35980000
#> [25,] 0.006629247  0.14026463 0.15121073 0.15008131 0.16557031 0.16189566
#>              V15       V16        V17       V18       V19           V2
#>  [1,] 0.17698175 0.1276309 0.16741309 0.2404766 0.3182653  0.020646392
#>  [2,] 0.45187949 0.4686049 0.62625756 0.7296039 0.8497131  0.046790036
#>  [3,] 0.19277166 0.3460446 0.52300837 0.5755973 0.7763953  0.048856183
#>  [4,] 0.43840885 0.5097215 0.20601103 0.1639421 0.2367541  0.079757337
#>  [5,] 0.13109979 0.1647276 0.19643562 0.2378516 0.1883759  0.034689450
#>  [6,] 0.25963193 0.2832731 0.25405309 0.1304014 0.2978155  0.048920051
#>  [7,] 0.45659001 0.4599566 0.28644539 0.3189968 0.4793377  0.085378772
#>  [8,] 0.31979161 0.3796728 0.39038381 0.4236680 0.5452382  0.041485505
#>  [9,] 0.29880277 0.2997742 0.25222910 0.2177162 0.2959011  0.030062703
#> [10,] 0.32981601 0.4178028 0.49383892 0.5243629 0.6016090  0.029681543
#> [11,] 0.74235850 0.7824114 0.85205711 0.9214875 0.9355567  0.061233199
#> [12,] 0.12501618 0.2026956 0.21026052 0.3133061 0.4752602  0.030488437
#> [13,] 0.58938262 0.5982824 0.72715858 0.7869070 0.8243702  0.105361917
#> [14,] 0.14959503 0.1814801 0.19909059 0.2498635 0.2285514  0.029112869
#> [15,] 0.24479938 0.2735367 0.33232460 0.3125188 0.2700472  0.023088759
#> [16,] 0.13568108 0.1038690 0.08865725 0.2042088 0.1359521  0.022974968
#> [17,] 0.10351958 0.1526530 0.19773067 0.1959274 0.1897540  0.018049421
#> [18,] 0.08236222 0.1350805 0.20020895 0.3484372 0.4180374  0.019720843
#> [19,] 0.24533914 0.2362522 0.22336319 0.2660258 0.2373397  0.007370052
#> [20,] 0.62519200 0.6687732 0.74453112 0.6257221 0.3556781  0.033761141
#> [21,] 0.18133313 0.2253910 0.21281408 0.3384317 0.6197317  0.037400343
#> [22,] 0.12167099 0.1733154 0.16193858 0.1297893 0.1759119  0.037823760
#> [23,] 0.06530000 0.0890000 0.12260000 0.1846000 0.3880000  0.014200000
#> [24,] 0.17130000 0.1136000 0.03490000 0.3796000 0.7401000  0.037500000
#> [25,] 0.22433342 0.2458606 0.28268100 0.2030921 0.2845410 -0.003169256
#>             V20        V21       V22       V23        V24       V25       V26
#>  [1,] 0.3971186 0.40010845 0.3635155 0.4347796 0.59239506 0.7046565 0.7741576
#>  [2,] 0.9655996 0.92345604 0.9442036 1.0044214 1.04064792 0.9005101 0.6063179
#>  [3,] 0.8373767 0.61275280 0.7010422 0.8559113 0.72407379 0.7826340 0.8012841
#>  [4,] 0.8577325 0.97210940 0.3829969 0.1283668 0.18792996 0.2006856 0.3599588
#>  [5,] 0.1175280 0.07977815 0.2105418 0.4198819 0.57426055 0.6962684 0.8321087
#>  [6,] 0.3814400 0.45824510 0.4396134 0.4671455 0.51382312 0.5275327 0.5500119
#>  [7,] 0.9009229 0.97778163 0.3989002 0.1097862 0.05965887 0.2425713 0.4399512
#>  [8,] 0.6996650 0.78161499 0.8399395 0.8391840 0.85005139 0.8124421 0.8258728
#>  [9,] 0.2908373 0.31833137 0.3016875 0.3703405 0.44874463 0.4253776 0.5961121
#> [10,] 0.7780497 0.93212926 1.0360920 0.9426101 0.70302136 0.4717321 0.2815775
#> [11,] 0.9024633 0.76212956 0.6124013 0.5715991 0.54265939 0.3953998 0.2729361
#> [12,] 0.5845517 0.68574779 0.7262777 0.7766349 0.86901061 0.9258212 0.9913031
#> [13,] 0.8545212 0.79646852 0.8725003 1.0646682 1.09946371 0.9783356 0.9262400
#> [14,] 0.2111547 0.31407244 0.6108443 0.8350660 0.89027958 0.9359926 0.9315671
#> [15,] 0.2934803 0.35121753 0.4283680 0.4991437 0.62154454 0.6735419 0.6775405
#> [16,] 0.1944796 0.27939015 0.1689603 0.1218014 0.19657364 0.1242102 0.1774562
#> [17,] 0.3142550 0.47873624 0.6934116 0.9071721 0.90239900 0.8607578 0.9010447
#> [18,] 0.5040818 0.56754064 0.5900696 0.6010054 0.60605230 0.4847280 0.5335758
#> [19,] 0.3328677 0.40402047 0.5013138 0.6807475 0.70989426 0.8297691 0.8860156
#> [20,] 0.3456378 0.51911575 0.7429712 0.8352279 0.82993223 0.7089713 0.6328536
#> [21,] 0.7750479 0.76660858 0.7303615 0.4866389 0.36589229 0.5575091 0.7691210
#> [22,] 0.1199696 0.18273507 0.0342451 0.2064580 0.18916317 0.0472457 0.1937948
#> [23,] 0.3658000 0.22970000 0.2610000 0.4193000 0.58480000 0.5643000 0.5448000
#> [24,] 0.9925000 0.98020000 0.8890000 0.6712000 0.42860000 0.3374000 0.7366000
#> [25,] 0.3849170 0.42668430 0.4022439 0.6110987 0.75631829 0.7733266 0.8200620
#>             V27        V28       V29          V3       V30       V31        V32
#>  [1,] 0.9396275 1.08123776 1.0410586 0.018276878 0.7503484 0.6520695 0.46248998
#>  [2,] 0.4245218 0.40003636 0.4578439 0.097316169 0.3951747 0.2994487 0.17412023
#>  [3,] 0.6773787 0.60910080 0.4082184 0.055953677 0.6532532 0.7358138 0.82864976
#>  [4,] 0.2831275 0.43518955 0.4196677 0.071122492 0.5923036 0.2400220 0.33102434
#>  [5,] 0.9316669 0.94724343 0.9101509 0.035641815 0.9376577 0.9081860 0.80836281
#>  [6,] 0.6054750 0.76808084 0.8505287 0.058769665 0.8999742 0.8840292 0.70845502
#>  [7,] 0.2617614 0.54763076 0.6530902 0.049955894 0.8356617 0.4663278 0.41785354
#>  [8,] 0.8760518 0.86897580 0.7550410 0.037237936 0.5642013 0.4712794 0.43373278
#>  [9,] 0.8124723 0.88642629 0.8959807 0.014638635 0.7693161 0.5198689 0.46553132
#> [10,] 0.1755729 0.09381873 0.2616649 0.012060568 0.3591784 0.2871216 0.51029221
#> [11,] 0.2104510 0.20765043 0.2080780 0.079307718 0.1888248 0.2262113 0.31893880
#> [12,] 0.9871479 0.94072703 0.8732317 0.036871775 0.7318748 0.5295000 0.31561205
#> [13,] 0.8853126 0.64696390 0.5929438 0.110884912 0.4630170 0.1961201 0.08963535
#> [14,] 0.9033230 0.95623412 0.8673443 0.045360944 0.6711678 0.7081022 0.66010394
#> [15,] 0.6982603 0.80300138 0.9147447 0.027301838 0.9188970 0.8275021 0.70853541
#> [16,] 0.2695567 0.43428605 0.5357583 0.017959768 0.6292888 0.6676866 0.63557121
#> [17,] 0.9086669 0.89276461 0.9188362 0.021810494 0.7779629 0.6468216 0.56358027
#> [18,] 0.6409483 0.77612240 0.9077651 0.023004547 0.8573063 0.7503122 0.62341316
#> [19,] 0.7281848 0.65920264 0.6161835 0.013541445 0.5433380 0.5256641 0.49733815
#> [20,] 0.7157486 0.67355819 0.4581646 0.055716420 0.4655126 0.4990230 0.36231193
#> [21,] 0.9351241 0.78926310 0.5592769 0.029799529 0.4082532 0.3937734 0.33921365
#> [22,] 0.3867737 0.66376358 0.6984874 0.033482637 0.6541144 0.6755304 0.55659778
#> [23,] 0.4772000 0.68970000 0.9797000 0.018100000 1.0000000 0.9546000 0.88350000
#> [24,] 0.9611000 0.73530000 0.4856000 0.048400000 0.1594000 0.3007000 0.40960000
#> [25,] 0.7797171 0.75403185 0.5851517 0.004147019 0.3777429 0.3139698 0.19708190
#>               V33       V34         V35        V36       V37       V38
#>  [1,]  0.33857608 0.2964187 0.151596495 0.08377303 0.1758095 0.1237261
#>  [2,]  0.23952356 0.3731522 0.535026583 0.48156282 0.3490433 0.4351546
#>  [3,]  0.76786836 0.6554771 0.418109039 0.18260837 0.1092301 0.3637969
#>  [4,]  0.54678080 0.7032534 0.754224316 0.80706903 0.7820640 0.8373715
#>  [5,]  0.70333912 0.6421246 0.655098179 0.69834502 0.6731475 0.5474039
#>  [6,]  0.48643979 0.4064681 0.344767614 0.23259646 0.2239628 0.3078855
#>  [7,]  0.55837622 0.7582842 0.810027609 0.74907119 0.7406645 0.8401986
#>  [8,]  0.39586987 0.2609870 0.155195696 0.23780349 0.2862374 0.3528797
#>  [9,]  0.50529495 0.3902149 0.192419420 0.15915047 0.1958897 0.1705929
#> [10,]  0.43562260 0.1665015 0.001552316 0.23538892 0.3706978 0.3650284
#> [11,]  0.29200203 0.2658361 0.316035185 0.20312423 0.2607118 0.3376187
#> [12,]  0.26553333 0.2455127 0.156067414 0.03946620 0.0859891 0.1278468
#> [13,] -0.01063682 0.1978597 0.367456735 0.32415572 0.2228278 0.3289883
#> [14,]  0.50602414 0.2092500 0.087778082 0.13672452 0.1660344 0.1502645
#> [15,]  0.67518862 0.6499314 0.645568983 0.61863767 0.5321430 0.3948360
#> [16,]  0.71296820 0.9158493 1.014759309 0.94582856 0.7793728 0.6110570
#> [17,]  0.55729987 0.6316010 0.642240959 0.53705993 0.3897364 0.2612899
#> [18,]  0.61432270 0.6238377 0.549436514 0.55562853 0.5183173 0.4389042
#> [19,]  0.62334369 0.7030142 0.707909816 0.58491246 0.3380278 0.2116223
#> [20,]  0.25567122 0.1144047 0.375210038 0.62996197 0.6815011 0.3436966
#> [21,]  0.33550905 0.2540190 0.262004653 0.30406695 0.4463785 0.4308982
#> [22,]  0.66575811 0.8309117 0.889116886 0.93086961 0.8000449 0.7258622
#> [23,]  0.76620000 0.6547000 0.544700000 0.45930000 0.4679000 0.1987000
#> [24,]  0.31700000 0.3305000 0.340800000 0.21860000 0.2463000 0.2726000
#> [25,]  0.16263773 0.2498421 0.263599715 0.21923306 0.1680928 0.1732800
#>             V39          V4        V40        V41       V42        V43
#>  [1,] 0.1450356  0.04836832 0.23521205 0.24937594 0.2992262 0.13697727
#>  [2,] 0.4878193  0.10439715 0.41577050 0.35557718 0.4331386 0.38503574
#>  [3,] 0.5839205  0.09010765 0.46979970 0.48690991 0.4821268 0.39175768
#>  [4,] 0.6830542  0.07790425 0.50054748 0.32731044 0.2150983 0.18669049
#>  [5,] 0.4391648  0.05443318 0.48987117 0.46805248 0.4171686 0.44196919
#>  [6,] 0.3862990  0.04638032 0.43549435 0.57885771 0.5269297 0.45240304
#>  [7,] 0.7629792  0.15504876 0.57378538 0.43406308 0.4662161 0.45057637
#>  [8,] 0.4302821  0.04622865 0.38645684 0.26962852 0.3450185 0.35414917
#>  [9,] 0.0978053  0.07077112 0.24402886 0.38392133 0.4533318 0.41534600
#> [10,] 0.2609607  0.04024923 0.16627577 0.13664641 0.1055027 0.09313525
#> [11,] 0.2520713  0.09319131 0.14526942 0.13510302 0.1973978 0.19779809
#> [12,] 0.1214726  0.03748754 0.11251381 0.07066923 0.1467132 0.21909575
#> [13,] 0.3786476  0.05958014 0.30980583 0.42195143 0.3457960 0.05745770
#> [14,] 0.1833700  0.06785601 0.08128429 0.17008251 0.1155964 0.13433028
#> [15,] 0.2852541  0.02721220 0.20327747 0.19404325 0.1721354 0.19893692
#> [16,] 0.5671687  0.03831434 0.53785211 0.46035984 0.4051110 0.24716486
#> [17,] 0.2004192  0.03791798 0.22802108 0.18137670 0.1371538 0.17863504
#> [18,] 0.3956279  0.03010149 0.31272306 0.25424934 0.1309460 0.10171835
#> [19,] 0.2402095  0.02649038 0.38990210 0.33971195 0.3694546 0.38542646
#> [20,] 0.2406188  0.03376135 0.42692805 0.45413161 0.3644718 0.22929705
#> [21,] 0.4048839  0.03668102 0.28107877 0.18358392 0.1593603 0.07725296
#> [22,] 0.7078188  0.02246831 0.72376273 0.65258341 0.5712024 0.41155232
#> [23,] 0.0699000  0.04060000 0.14930000 0.17130000 0.1654000 0.26000000
#> [24,] 0.1680000  0.04750000 0.27920000 0.25580000 0.1740000 0.21210000
#> [25,] 0.1968570 -0.01009917 0.12634718 0.10621972 0.1443347 0.16367220
#>              V44        V45         V46        V47        V48         V49
#>  [1,] 0.10309912 0.19570632 0.177289511 0.22803086 0.14129906 0.055334237
#>  [2,] 0.27356267 0.23966592 0.189871842 0.12714938 0.05336261 0.042728572
#>  [3,] 0.32755522 0.43279381 0.450158481 0.27027066 0.21155132 0.122423105
#>  [4,] 0.21117052 0.38571804 0.276222308 0.14473017 0.15555201 0.099361549
#>  [5,] 0.50127214 0.45394040 0.328163970 0.23147716 0.18295542 0.101376817
#>  [6,] 0.44032114 0.58252943 0.541777899 0.38084187 0.24023789 0.131909280
#>  [7,] 0.39196878 0.49302768 0.395879524 0.08135016 0.12939910 0.064769313
#>  [8,] 0.33922044 0.29658388 0.138593224 0.10381214 0.12228234 0.072507618
#>  [9,] 0.22687489 0.24297838 0.245050256 0.18966466 0.16840315 0.109573282
#> [10,] 0.08804702 0.03753115 0.009671785 0.04070340 0.03884094 0.011155627
#> [11,] 0.12439044 0.06888326 0.032319605 0.03226921 0.05276782 0.029725005
#> [12,] 0.20293358 0.16912024 0.180873315 0.15486557 0.09786480 0.056219672
#> [13,] 0.18372014 0.19355293 0.180424307 0.11444584 0.09277501 0.043960831
#> [14,] 0.15198269 0.09322098 0.047124302 0.10902286 0.14204238 0.078415126
#> [15,] 0.17612153 0.13715979 0.101538288 0.06339144 0.04520892 0.022361566
#> [16,] 0.17899213 0.14381252 0.102592520 0.10289852 0.08956797 0.039098146
#> [17,] 0.11528503 0.04484184 0.057488590 0.04498131 0.03667792 0.007220907
#> [18,] 0.11033168 0.09683302 0.057480985 0.04198906 0.03931416 0.023046154
#> [19,] 0.27323491 0.07857163 0.030372086 0.04600307 0.04732710 0.007420272
#> [20,] 0.14768707 0.15527093 0.156246116 0.14281718 0.11805564 0.080612888
#> [21,] 0.05974131 0.07955373 0.077902954 0.08301504 0.04924915 0.025126736
#> [22,] 0.22850382 0.20408405 0.169071957 0.15379636 0.12117463 0.086402704
#> [23,] 0.38460000 0.37540000 0.241400000 0.10770000 0.02240000 0.015500000
#> [24,] 0.10990000 0.09850000 0.127100000 0.14590000 0.11640000 0.077700000
#> [25,] 0.13457834 0.08557213 0.059760431 0.01498626 0.01658467 0.012024878
#>               V5         V50         V51         V52          V53         V54
#>  [1,] 0.11199218 0.010073394 0.025894146 0.014908900 1.235104e-02 0.004648054
#>  [2,] 0.16437282 0.012193251 0.005910561 0.011044769 1.854159e-02 0.016138623
#>  [3,] 0.18963258 0.068766984 0.018802662 0.016818519 2.551160e-02 0.016987332
#>  [4,] 0.05793796 0.039369999 0.026914632 0.012432619 1.072179e-02 0.030724538
#>  [5,] 0.07568361 0.029604700 0.016339944 0.013618938 8.513881e-03 0.005530378
#>  [6,] 0.06116127 0.047178145 0.029074934 0.022582861 1.511048e-02 0.014396764
#>  [7,] 0.13161181 0.038802253 0.017937522 0.033836020 3.444984e-02 0.023950088
#>  [8,] 0.06320792 0.015391969 0.021339489 0.015446816 1.077959e-02 0.010211434
#>  [9,] 0.08362257 0.020704309 0.014055513 0.012581000 3.985716e-03 0.003472376
#> [10,] 0.04204942 0.012788312 0.013895395 0.010779583 1.944567e-05 0.006859817
#> [11,] 0.11370727 0.013454005 0.019021908 0.023679624 1.938074e-02 0.012381608
#> [12,] 0.05625856 0.015508406 0.013794433 0.012248303 5.395646e-03 0.002871296
#> [13,] 0.16296754 0.022649473 0.025211048 0.020527576 2.561305e-02 0.035068500
#> [14,] 0.08685302 0.021107584 0.011419151 0.011026952 9.532304e-03 0.014028854
#> [15,] 0.03353325 0.017640826 0.008895831 0.009335250 8.729670e-03 0.009917927
#> [16,] 0.02466765 0.020805147 0.023106035 0.003948162 4.571000e-03 0.007429041
#> [17,] 0.06081969 0.007201523 0.003243410 0.003426696 4.910408e-03 0.005832772
#> [18,] 0.04956496 0.012074192 0.003114772 0.006082605 6.342266e-03 0.006109111
#> [19,] 0.05964936 0.005853478 0.016880661 0.002732806 7.049589e-03 0.009545916
#> [20,] 0.03609266 0.019829305 0.020552100 0.009664749 7.937994e-03 0.009365976
#> [21,] 0.03196740 0.019991432 0.016633464 0.012168654 2.079875e-02 0.022379269
#> [22,] 0.05935694 0.030658359 0.012473491 0.009892431 1.630365e-02 0.015243952
#> [23,] 0.03910000 0.018700000 0.012500000 0.002800000 6.700000e-03 0.012000000
#> [24,] 0.06470000 0.043900000 0.006100000 0.014500000 1.280000e-02 0.014500000
#> [25,] 0.01228391 0.006980089 0.010406350 0.015944890 7.720153e-03 0.004109590
#>               V55         V56          V57         V58           V59
#>  [1,] 0.001247843 0.005417172 0.0009909957 0.001917315 -0.0002664656
#>  [2,] 0.006898918 0.015047146 0.0163856152 0.017339317  0.0202497463
#>  [3,] 0.032817524 0.027687232 0.0259874149 0.032832550  0.0190180702
#>  [4,] 0.019181279 0.021969636 0.0139622806 0.018408788  0.0204216153
#>  [5,] 0.005441077 0.007573510 0.0051866465 0.005328667  0.0039346929
#>  [6,] 0.010659595 0.010471277 0.0083543428 0.010653130  0.0112576278
#>  [7,] 0.016188936 0.022792015 0.0131231290 0.010693341  0.0064331912
#>  [8,] 0.009621440 0.007507105 0.0055537349 0.005446071  0.0053004253
#>  [9,] 0.009690873 0.010475219 0.0048483960 0.012968129  0.0137635087
#> [10,] 0.002851164 0.004733510 0.0065470280 0.006563635  0.0004136460
#> [11,] 0.021313585 0.021265606 0.0076258166 0.014828089  0.0148483119
#> [12,] 0.004828507 0.003631691 0.0045583485 0.002506515  0.0019618570
#> [13,] 0.028688533 0.013688424 0.0153376119 0.019593150  0.0040442804
#> [14,] 0.006340486 0.008317680 0.0055470896 0.006087457  0.0091931112
#> [15,] 0.009307502 0.005353106 0.0073488845 0.004360862  0.0099903121
#> [16,] 0.011508681 0.009700929 0.0060951627 0.005930185  0.0067956726
#> [17,] 0.005822602 0.005289506 0.0093076578 0.008772819  0.0048586997
#> [18,] 0.004989791 0.004867991 0.0061210334 0.004708374  0.0056478454
#> [19,] 0.002904210 0.009842863 0.0101813395 0.002888465  0.0056312677
#> [20,] 0.007299517 0.005500216 0.0130102190 0.007139450  0.0075594962
#> [21,] 0.011860567 0.010944858 0.0070871499 0.013325207  0.0039491944
#> [22,] 0.010849757 0.007741084 0.0102442257 0.010730307  0.0114517158
#> [23,] 0.001200000 0.002200000 0.0058000000 0.004200000  0.0067000000
#> [24,] 0.005800000 0.004900000 0.0065000000 0.009300000  0.0059000000
#> [25,] 0.009118510 0.006810677 0.0055087187 0.007765965  0.0043897975
#>                V6         V60         V7         V8         V9
#>  [1,] 0.147475853 0.000689209 0.14500222 0.12215638 0.16027706
#>  [2,] 0.117538760 0.011427211 0.14452646 0.06835843 0.02826227
#>  [3,] 0.125373263 0.007747137 0.16240760 0.11868845 0.24311945
#>  [4,] 0.120270292 0.008003693 0.15893450 0.43479204 0.59255781
#>  [5,] 0.108367681 0.003717815 0.12907777 0.15745519 0.27756225
#>  [6,] 0.121530270 0.007225945 0.12255453 0.14565575 0.24159303
#>  [7,] 0.061289234 0.012310813 0.10735729 0.30852010 0.59632310
#>  [8,] 0.102606012 0.003540722 0.10776662 0.15478611 0.23300415
#>  [9,] 0.070732011 0.005524214 0.09219079 0.17455846 0.20368690
#> [10,] 0.071196742 0.002245889 0.13776432 0.14544448 0.20672176
#> [11,] 0.090640573 0.009924770 0.09439902 0.10506148 0.13143652
#> [12,] 0.082198125 0.004860298 0.12510660 0.16849260 0.22646515
#> [13,] 0.234115855 0.003218153 0.23901779 0.32461772 0.43513136
#> [14,] 0.124918821 0.008314082 0.13253218 0.14055448 0.14883432
#> [15,] 0.068953346 0.005723341 0.07835590 0.07121932 0.09105777
#> [16,] 0.076385250 0.003490707 0.07118728 0.06321093 0.02244628
#> [17,] 0.091097111 0.004028439 0.08510148 0.10008399 0.10347232
#> [18,] 0.104961880 0.004130212 0.07594660 0.11209036 0.12580871
#> [19,] 0.100720992 0.005661820 0.10703991 0.08121093 0.05470125
#> [20,] 0.084541564 0.006979115 0.09785453 0.21895471 0.26572245
#> [21,] 0.006295133 0.004615617 0.02162192 0.05164506 0.11080140
#> [22,] 0.102484716 0.005585063 0.11762725 0.10073449 0.00848801
#> [23,] 0.024900000 0.001200000 0.08920000 0.09730000 0.08400000
#> [24,] 0.059100000 0.002200000 0.07530000 0.00980000 0.06840000
#> [25,] 0.068837718 0.002626032 0.07293019 0.10867637 0.09528320
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623