Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

r format_bib("kohonen_1990", "kohonen_1995")

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10        V11        V12        V13        V14
#>  [1,] 0.028972836 0.207873614 0.26451580 0.28395117 0.26396963 0.19885023
#>  [2,] 0.027262317 0.290660044 0.34394125 0.35454630 0.35196163 0.33817489
#>  [3,] 0.084236097 0.205964242 0.26875990 0.25534153 0.29328939 0.34581748
#>  [4,] 0.112919681 0.214105790 0.38850709 0.47137684 0.64368452 0.75887553
#>  [5,] 0.025092260 0.326687408 0.25718430 0.29347027 0.25966427 0.21077859
#>  [6,] 0.037028407 0.298810143 0.34261363 0.41181074 0.37276571 0.37749632
#>  [7,] 0.026544778 0.315541177 0.26706569 0.27140183 0.23197221 0.21577237
#>  [8,] 0.034885725 0.147816573 0.22481116 0.28080788 0.28457834 0.29450734
#>  [9,] 0.065372165 0.349575400 0.51399828 0.47753061 0.48082942 0.48578925
#> [10,] 0.012069921 0.224854863 0.11456204 0.04541885 0.21880566 0.33023696
#> [11,] 0.082625044 0.274090102 0.31343328 0.29282392 0.32913986 0.47191537
#> [12,] 0.025966693 0.385034397 0.41862910 0.42736926 0.51953352 0.44339782
#> [13,] 0.003785689 0.366190552 0.34096719 0.33166413 0.23067446 0.11885440
#> [14,] 0.016711675 0.235626186 0.28703709 0.31116299 0.28676189 0.23500151
#> [15,] 0.026384601 0.067057771 0.02834928 0.04841091 0.08246901 0.09195430
#> [16,] 0.007082935 0.100674267 0.10535682 0.05110662 0.02541670 0.05709727
#> [17,] 0.021671414 0.132914130 0.09306503 0.02930737 0.14745288 0.29505352
#> [18,] 0.029713373 0.280738394 0.22418724 0.07435696 0.19758739 0.45348771
#> [19,] 0.017380075 0.011404230 0.14223560 0.36703270 0.27590508 0.21831158
#> [20,] 0.023447911 0.121879962 0.16788581 0.21409539 0.31392601 0.40766529
#> [21,] 0.005389486 0.189735283 0.09671215 0.11653526 0.27782868 0.26409605
#> [22,] 0.017094058 0.006996116 0.04785426 0.20773205 0.17881704 0.07779837
#> [23,] 0.022925972 0.098164829 0.10356566 0.16247228 0.23561066 0.37285090
#> [24,] 0.029957362 0.276005237 0.31447729 0.38239653 0.40929292 0.47786437
#> [25,] 0.013600682 0.137625881 0.14441390 0.14890093 0.16786703 0.18354239
#>              V15        V16        V17       V18       V19          V2
#>  [1,] 0.17207873 0.18205736 0.20304407 0.2482143 0.2709175 0.034717954
#>  [2,] 0.33157962 0.34699553 0.41068854 0.5226780 0.6569542 0.030477530
#>  [3,] 0.37954740 0.33202572 0.38635538 0.2137046 0.2835109 0.081785364
#>  [4,] 0.79403557 0.85686048 0.91280406 0.8966317 0.9428301 0.106818612
#>  [5,] 0.11097391 0.15889475 0.16833987 0.2333548 0.1820662 0.031761529
#>  [6,] 0.30506356 0.31199808 0.26893628 0.2627055 0.2667245 0.028882973
#>  [7,] 0.12644668 0.11389241 0.16315955 0.2091277 0.1644332 0.045342527
#>  [8,] 0.32719818 0.44761741 0.61095159 0.6993506 0.7967252 0.038896007
#>  [9,] 0.62745987 0.82866526 0.96047511 0.9378159 0.8049825 0.119310381
#> [10,] 0.47473309 0.64126519 0.70177471 0.7504556 0.8340154 0.009465064
#> [11,] 0.47713926 0.32213897 0.26154340 0.2910518 0.3343136 0.111538122
#> [12,] 0.29578304 0.33217750 0.36212967 0.3299731 0.5142189 0.040482152
#> [13,] 0.21017328 0.22864019 0.15567094 0.1326030 0.1044307 0.004120602
#> [14,] 0.17840046 0.20117755 0.23908914 0.3667184 0.5872233 0.043856141
#> [15,] 0.13175886 0.06799225 0.12703217 0.2287170 0.3313891 0.062749944
#> [16,] 0.11669302 0.09452299 0.08681612 0.1660925 0.2501132 0.004499489
#> [17,] 0.55956562 0.80080006 0.85014647 0.8080462 0.8869902 0.016528151
#> [18,] 0.63371694 0.78156927 0.83054441 0.7390960 0.5910839 0.050766661
#> [19,] 0.08790084 0.08385320 0.16365624 0.4020042 0.2071579 0.055336766
#> [20,] 0.39439058 0.36051076 0.30525216 0.3352490 0.3602540 0.027905204
#> [21,] 0.16599315 0.27464716 0.35031424 0.3787472 0.5458964 0.043044468
#> [22,] 0.08686185 0.15367183 0.24187254 0.2362683 0.2074734 0.024617985
#> [23,] 0.47198657 0.53612169 0.62386187 0.6878390 0.7994687 0.028315397
#> [24,] 0.63791460 0.73249122 0.87296441 0.9143203 0.8914527 0.039495164
#> [25,] 0.20041563 0.26501876 0.35259648 0.3704827 0.3402886 0.018631308
#>              V20        V21       V22        V23       V24       V25       V26
#>  [1,] 0.28278114 0.38517838 0.5150573 0.64553034 0.7426699 0.8142716 0.8615361
#>  [2,] 0.73683383 0.80616870 0.8628062 0.89892449 0.9552790 0.9323463 0.9226614
#>  [3,] 0.41512340 0.44299008 0.3891495 0.42958375 0.3407268 0.4342034 0.4573116
#>  [4,] 0.84190236 0.84514634 0.8586322 0.83953487 0.7235744 0.5423928 0.5424251
#>  [5,] 0.15539731 0.03337803 0.1395427 0.36520102 0.5194860 0.6481526 0.7705169
#>  [6,] 0.51428684 0.62831099 0.5629398 0.54469580 0.6300715 0.6017886 0.8096690
#>  [7,] 0.06959731 0.04601764 0.2166727 0.47320824 0.6325946 0.7501738 0.8996062
#>  [8,] 0.80988284 0.84033851 0.8995938 0.95040333 0.9073032 0.7733388 0.6527111
#>  [9,] 0.73596342 0.73020879 0.4072412 0.05204733 0.0374303 0.2533355 0.2544828
#> [10,] 0.85420332 0.94795361 0.9779364 0.89083804 0.6402490 0.4725830 0.5178863
#> [11,] 0.62783341 0.73238502 0.4770039 0.38172816 0.3373555 0.2009503 0.4840067
#> [12,] 0.73583818 0.75078089 0.7250524 0.66527229 0.6495389 0.6924652 0.8476458
#> [13,] 0.07771691 0.14017702 0.3153538 0.51964812 0.6711751 0.7880106 0.8938586
#> [14,] 0.69669240 0.79087901 0.7957236 0.81098724 0.9255767 0.9435940 0.9627846
#> [15,] 0.31337424 0.19378397 0.2242048 0.34268470 0.4477793 0.4819619 0.2646602
#> [16,] 0.27753862 0.32519401 0.4762868 0.67054822 0.8519855 0.9682644 0.8543155
#> [17,] 0.88867837 0.96951762 0.9708465 0.96778414 0.9673525 0.9095208 0.9664024
#> [18,] 0.50886121 0.63080141 0.7556172 0.84836464 0.9624489 0.8895869 0.7737368
#> [19,] 0.53926652 0.70214364 0.7103022 0.76170690 0.8249233 0.9918803 0.8858046
#> [20,] 0.45470569 0.52624095 0.5214945 0.54513661 0.7244689 0.8993200 0.8966422
#> [21,] 0.63885738 0.69866102 0.6496832 0.58751799 0.4342160 0.6136682 0.8636511
#> [22,] 0.11547715 0.09098633 0.1261681 0.19143793 0.1618565 0.1510037 0.2916166
#> [23,] 0.88580319 0.92509311 0.8243452 0.82665241 0.8800716 0.8327738 0.8191980
#> [24,] 0.82643948 0.76615668 0.7506005 0.62182259 0.5466723 0.5869666 0.6357971
#> [25,] 0.37452464 0.42793427 0.5552429 0.68427526 0.8070681 0.8352629 0.8349032
#>             V27       V28          V29          V3       V30       V31
#>  [1,] 0.9156348 0.9712167  0.883063865  0.03299679 0.7015829 0.6463442
#>  [2,] 0.9262269 0.8675504  0.725321274  0.03669950 0.5577930 0.4267290
#>  [3,] 0.4811173 0.7421559  0.774330825  0.07480827 0.7053595 0.7449868
#>  [4,] 0.4698578 0.5013751  0.476663633  0.09714581 0.3213922 0.2275803
#>  [5,] 0.9051322 0.9491997  0.898358011  0.02117626 0.9301705 0.9376661
#>  [6,] 0.9088798 0.8829183  0.803323843  0.06265573 0.7193211 0.6511098
#>  [7,] 1.0068658 1.0086275  0.933728334  0.05627228 0.9363026 0.8917533
#>  [8,] 0.5138515 0.3888031  0.386415793  0.05600533 0.3880602 0.3411026
#>  [9,] 0.1949009 0.1737667 -0.002678514  0.16039375 0.2462039 0.1792401
#> [10,] 0.6522736 0.8125664  0.811212548  0.02895925 0.6064975 0.3175748
#> [11,] 0.5655507 0.4601039  0.432083081  0.14307227 0.5455182 0.3925072
#> [12,] 0.9432385 0.9054183  0.766498369  0.05310073 0.6075967 0.5467690
#> [13,] 0.9426249 0.9573909  0.904727678  0.00602878 0.9153837 0.8860650
#> [14,] 0.9881704 0.9673762  0.858979595  0.03673442 0.6388212 0.4102579
#> [15,] 0.0871107 0.2832995  0.565811854  0.04013234 0.8294668 0.8182859
#> [16,] 0.7522499 0.7185401  0.578948842 -0.01385513 0.3764916 0.2098083
#> [17,] 0.9678279 0.7796750  0.474803555  0.04014675 0.3862728 0.4407328
#> [18,] 0.7895212 0.6400135  0.419074292  0.07067725 0.4189403 0.4709701
#> [19,] 0.7025700 0.5401725  0.506159842  0.06751202 0.5155967 0.6004452
#> [20,] 0.8494066 0.8655923  0.849470976  0.01857491 0.7305599 0.6081147
#> [21,] 0.8763694 0.5080015  0.231462906  0.06940333 0.1479640 0.3357062
#> [22,] 0.4809028 0.6335115  0.750309428  0.03623823 0.7150125 0.7928416
#> [23,] 0.7274444 0.4956044  0.350699193  0.02238912 0.2852231 0.2713233
#> [24,] 0.6034347 0.6310876  0.711634321  0.05195443 0.7315514 0.5935166
#> [25,] 0.7990475 0.8265068  0.851348171  0.03064377 0.7855567 0.7241574
#>             V32       V33       V34        V35        V36        V37       V38
#>  [1,] 0.5501249 0.4384261 0.2665933 0.14889751 0.12971961 0.18071278 0.1707818
#>  [2,] 0.3418098 0.2961711 0.2502364 0.23940939 0.23255100 0.23179597 0.2427640
#>  [3,] 0.7342523 0.6100858 0.6227939 0.64928451 0.40155285 0.15328693 0.3327132
#>  [4,] 0.1866040 0.2072805 0.2238623 0.25263608 0.09998885 0.15959978 0.2261398
#>  [5,] 0.8905716 0.7541075 0.6453814 0.65594349 0.71481982 0.75652110 0.6309532
#>  [6,] 0.7210555 0.5736504 0.2800558 0.06318922 0.16364702 0.26328678 0.3997061
#>  [7,] 0.7735707 0.6846868 0.6368437 0.63516936 0.66088887 0.59762457 0.4492216
#>  [8,] 0.2974274 0.4206940 0.4910977 0.47363391 0.38040183 0.35449047 0.3612400
#>  [9,] 0.1924726 0.2699370 0.2355286 0.33703884 0.10915606 0.06753555 0.2317921
#> [10,] 0.3280015 0.3491284 0.2366667 0.15607390 0.09116376 0.09506131 0.1174500
#> [11,] 0.2327843 0.2714915 0.3033721 0.27003668 0.35102868 0.36109512 0.3744617
#> [12,] 0.6487893 0.5613653 0.3512751 0.10779886 0.23565860 0.26056662 0.3691950
#> [13,] 0.8147678 0.7188220 0.6626982 0.62375117 0.66528367 0.60232989 0.4732775
#> [14,] 0.3100367 0.2570629 0.2128991 0.18639905 0.09443295 0.07259942 0.0932993
#> [15,] 0.7352131 0.8002039 0.9448699 0.99474419 0.89057272 0.54242035 0.2838332
#> [16,] 0.2125105 0.2241193 0.2583909 0.19729439 0.20371980 0.26181266 0.2601220
#> [17,] 0.3940934 0.3241129 0.3006612 0.22988896 0.10196532 0.25993601 0.1895287
#> [18,] 0.3409131 0.2697676 0.2202785 0.45297755 0.53322373 0.53189589 0.3465234
#> [19,] 0.5517638 0.6218782 0.5301652 0.53610450 0.53898049 0.47786052 0.2528859
#> [20,] 0.4597876 0.3454283 0.2546731 0.20432267 0.28192834 0.18490947 0.2570878
#> [21,] 0.5271883 0.4362653 0.3147432 0.30881521 0.41202416 0.53602024 0.3661695
#> [22,] 0.6602481 0.6592898 0.7797675 0.81516596 0.86172398 0.78757288 0.6936364
#> [23,] 0.1649433 0.1885323 0.2266999 0.18957457 0.12148583 0.14753738 0.1695124
#> [24,] 0.3909487 0.3451828 0.3624481 0.26923916 0.19971871 0.30014293 0.2534699
#> [25,] 0.6965898 0.6878827 0.6928462 0.65744957 0.60994585 0.52106516 0.4001507
#>              V39          V4        V40        V41        V42       V43
#>  [1,] 0.19635559  0.04314957 0.18273029 0.22060329 0.21556773 0.1834278
#>  [2,] 0.22584981  0.04452190 0.22216568 0.20090332 0.17974319 0.2006570
#>  [3,] 0.29541802  0.09330828 0.14176303 0.50758704 0.70537831 0.5677926
#>  [4,] 0.28623930  0.10233932 0.23680854 0.10988784 0.23242785 0.3219133
#>  [5,] 0.54106119  0.03512636 0.56235410 0.50572021 0.42749687 0.4539688
#>  [6,] 0.51548975  0.01673513 0.45281853 0.52512373 0.66238324 0.5967537
#>  [7,] 0.29975271  0.07123081 0.41949180 0.44763331 0.42445678 0.4408766
#>  [8,] 0.36749073  0.06842354 0.40346705 0.41405470 0.42662089 0.3304271
#>  [9,] 0.20514391  0.13802236 0.02265656 0.11741729 0.17439813 0.1430797
#> [10,] 0.10759369  0.05389547 0.15600141 0.21927511 0.23858894 0.1489681
#> [11,] 0.45400137  0.17599168 0.37336627 0.24720921 0.28879928 0.2718974
#> [12,] 0.57565819  0.04749514 0.44484199 0.28540351 0.41747394 0.4142325
#> [13,] 0.39979938  0.05723063 0.42351051 0.46600272 0.45561150 0.4467907
#> [14,] 0.12225612  0.02698505 0.12869375 0.05607258 0.09505286 0.1866411
#> [15,] 0.21700950  0.01667891 0.23119320 0.33511017 0.38651917 0.1545717
#> [16,] 0.28073084 -0.01659059 0.27465957 0.26478375 0.20473604 0.1482649
#> [17,] 0.07773041  0.04062880 0.09871135 0.06850770 0.03322386 0.0848079
#> [18,] 0.19747292  0.05635298 0.28441750 0.26621776 0.24662233 0.2075899
#> [19,] 0.01554456  0.05468012 0.34271566 0.45113987 0.61719644 0.5161424
#> [20,] 0.33381015  0.02067324 0.23942483 0.12072972 0.21719961 0.2016873
#> [21,] 0.47783227  0.05090953 0.48952908 0.36334841 0.27085516 0.2368922
#> [22,] 0.63578121  0.01073321 0.65645113 0.59207025 0.51889802 0.3849069
#> [23,] 0.18936583  0.03634772 0.16530765 0.11992754 0.09479385 0.1135024
#> [24,] 0.20107840  0.04452451 0.25499411 0.22520653 0.21939533 0.1484692
#> [25,] 0.28707550  0.04005693 0.25121166 0.22112647 0.15514883 0.1350110
#>             V44        V45        V46        V47        V48        V49
#>  [1,] 0.1344893 0.11676604 0.11770605 0.14277903 0.12799606 0.06578703
#>  [2,] 0.1991975 0.16527193 0.12752629 0.09218386 0.06388790 0.04136335
#>  [3,] 0.4867259 0.58676636 0.64316416 0.42948937 0.24957443 0.11680054
#>  [4,] 0.1748861 0.10859510 0.12673762 0.08805863 0.07582025 0.06642893
#>  [5,] 0.5400118 0.52823444 0.39743405 0.29993730 0.23634065 0.12562144
#>  [6,] 0.5447122 0.59960989 0.41281531 0.28876224 0.23967533 0.13416108
#>  [7,] 0.4805100 0.41318554 0.30966589 0.21931815 0.16476200 0.09418651
#>  [8,] 0.2372619 0.16967838 0.12018169 0.11998031 0.07775738 0.04961336
#>  [9,] 0.1454371 0.06667518 0.01955847 0.08983843 0.08371340 0.01859226
#> [10,] 0.1984088 0.19157799 0.18669527 0.12175972 0.04941455 0.03998668
#> [11,] 0.1805709 0.21044087 0.17761843 0.12659585 0.11214392 0.07617898
#> [12,] 0.4587367 0.50459206 0.25697521 0.16085362 0.18860276 0.10319304
#> [13,] 0.5151087 0.48990068 0.32892896 0.23244204 0.19065801 0.10497211
#> [14,] 0.2126091 0.19599463 0.19947533 0.15542152 0.09707545 0.04780071
#> [15,] 0.2195255 0.14778168 0.13762519 0.16266803 0.11450257 0.05728065
#> [16,] 0.1710440 0.15869357 0.13245437 0.04064151 0.02892317 0.02571571
#> [17,] 0.1294804 0.11907000 0.03639063 0.05011966 0.04347138 0.02027041
#> [18,] 0.1860897 0.18678024 0.13632892 0.11989606 0.11026624 0.06906467
#> [19,] 0.2748138 0.17997528 0.15084635 0.08671066 0.07760455 0.01989023
#> [20,] 0.1631484 0.10306646 0.07593247 0.08916245 0.06105429 0.02678509
#> [21,] 0.1640663 0.02129437 0.01464797 0.04607316 0.03130852 0.03981550
#> [22,] 0.2271482 0.13603833 0.11851321 0.12761718 0.08959367 0.05336108
#> [23,] 0.1283395 0.09561151 0.06084402 0.07122121 0.04687543 0.02921366
#> [24,] 0.1127751 0.17778755 0.17991780 0.16058389 0.10305733 0.05416753
#> [25,] 0.1027588 0.08089029 0.06947331 0.04082710 0.03232910 0.01302680
#>                V5         V50          V51           V52         V53
#>  [1,] 0.056820229 0.016097777 1.741441e-02  1.401012e-02 0.008100699
#>  [2,] 0.086572190 0.017608805 1.484974e-02  1.470859e-02 0.010403260
#>  [3,] 0.065259875 0.043075690 7.319586e-02  5.970561e-02 0.030681131
#>  [4,] 0.142102033 0.032455878 2.526254e-02  1.569791e-02 0.016291108
#>  [5,] 0.056841595 0.038181317 2.344013e-02  1.965251e-02 0.008154660
#>  [6,] 0.097468722 0.039100771 2.875835e-02  1.510772e-02 0.007497602
#>  [7,] 0.083775098 0.026793001 1.252666e-02  9.393677e-03 0.006179042
#>  [8,] 0.099482094 0.015591277 9.936723e-03  9.054816e-03 0.013664548
#>  [9,] 0.148397131 0.009964667 2.735521e-02  1.357739e-02 0.026551518
#> [10,] 0.089587687 0.010575150 4.361271e-03  7.730002e-03 0.004459342
#> [11,] 0.155504501 0.030025739 1.782235e-02  2.020187e-02 0.017107459
#> [12,] 0.094789228 0.030635936 3.178589e-02  1.825644e-02 0.008199429
#> [13,] 0.093424850 0.024563897 1.283922e-02  9.242928e-03 0.013095835
#> [14,] 0.049035467 0.013708863 1.915013e-02  1.402190e-02 0.007542602
#> [15,] 0.049393278 0.012848647 1.695324e-02  2.952800e-03 0.005233458
#> [16,] 0.005509764 0.005988622 8.648078e-05  2.333133e-03 0.005149076
#> [17,] 0.067886741 0.008555812 1.549510e-02  9.363385e-03 0.012483080
#> [18,] 0.070918315 0.018560151 1.608301e-02  1.339861e-02 0.006894475
#> [19,] 0.128599588 0.008874678 4.718631e-03 -6.490601e-05 0.010468913
#> [20,] 0.040227529 0.016457400 5.193330e-03  6.035569e-03 0.004708751
#> [21,] 0.055597140 0.017615144 8.455056e-03  1.050144e-02 0.007939171
#> [22,] 0.021444606 0.023865467 1.185152e-02  8.445582e-03 0.017842281
#> [23,] 0.059673100 0.020982070 1.304844e-02  1.078377e-02 0.009823923
#> [24,] 0.059609897 0.023758299 9.874658e-03  8.686902e-03 0.005966200
#> [25,] 0.051731996 0.012862450 1.153543e-02  1.031820e-02 0.010600949
#>                V54           V55         V56          V57         V58
#>  [1,] 0.0095425331  6.507271e-03 0.008449836 7.017664e-03 0.008300426
#>  [2,] 0.0121704530  7.304496e-03 0.009544180 7.100947e-03 0.007315059
#>  [3,] 0.0298380000  2.389327e-02 0.010771638 1.715156e-02 0.016242609
#>  [4,] 0.0290159234  1.392093e-02 0.006598271 2.036905e-03 0.012132452
#>  [5,] 0.0081754016  6.973778e-03 0.007014125 7.629752e-03 0.002985354
#>  [6,] 0.0094381556  1.000679e-02 0.006464430 5.664963e-03 0.003807796
#>  [7,] 0.0014283559  4.146241e-03 0.008128527 2.843873e-03 0.006598100
#>  [8,] 0.0127576964  5.213319e-03 0.010257571 9.218897e-03 0.009266992
#>  [9,] 0.0115351445  1.499585e-02 0.017394342 2.103372e-02 0.006682094
#> [10,] 0.0008207796 -7.790553e-05 0.003494351 2.104706e-03 0.007298436
#> [11,] 0.0129372257  1.156529e-02 0.007961417 1.304222e-02 0.014169099
#> [12,] 0.0112178547  1.722703e-02 0.009918444 9.302525e-03 0.010583541
#> [13,] 0.0065196819  3.458587e-03 0.006977444 5.755432e-03 0.004309531
#> [14,] 0.0058913060  3.728878e-03 0.003654100 3.660659e-03 0.003109363
#> [15,] 0.0077108891  7.395787e-03 0.007661815 1.069342e-02 0.015867829
#> [16,] 0.0080206572 -5.464477e-04 0.000247701 8.527982e-05 0.002767538
#> [17,] 0.0083380492  2.589331e-03 0.007622777 4.970712e-03 0.003359300
#> [18,] 0.0121987119  8.375427e-03 0.004936390 9.385614e-03 0.006492729
#> [19,] 0.0071657773  1.169378e-02 0.009858152 1.004686e-02 0.006377281
#> [20,] 0.0046997412  3.857472e-03 0.007774812 5.375773e-03 0.003594898
#> [21,] 0.0168469103  1.631394e-02 0.005833622 1.067501e-02 0.017415694
#> [22,] 0.0084456419  8.140747e-03 0.005349812 4.298549e-03 0.011436708
#> [23,] 0.0040992874  8.135714e-03 0.005746933 5.114159e-03 0.003431858
#> [24,] 0.0095965046  1.218267e-02 0.010106952 1.160895e-02 0.007707851
#> [25,] 0.0059607455  7.363766e-03 0.007170376 8.244677e-03 0.006400389
#>                 V59           V6          V60         V7          V8
#>  [1,]  0.0073739340  0.112263595 0.0044396108 0.12863764 0.116526674
#>  [2,]  0.0068927788  0.124862951 0.0052618542 0.14926123 0.191838216
#>  [3,]  0.0192398140  0.112962475 0.0098648110 0.07648808 0.154549295
#>  [4,]  0.0143615513  0.155464148 0.0078358169 0.18732481 0.185896974
#>  [5,]  0.0026293500  0.076692293 0.0012053667 0.10836862 0.122582271
#>  [6,]  0.0050676029  0.111460463 0.0023646040 0.10765638 0.134467386
#>  [7,]  0.0046160991  0.131802166 0.0049008487 0.13401686 0.181213308
#>  [8,]  0.0130790381  0.107372013 0.0084641718 0.10440529 0.100362661
#>  [9,]  0.0146112123  0.081219798 0.0145827374 0.06379738 0.038456017
#> [10,]  0.0024546075  0.117294592 0.0028798950 0.13861087 0.173962708
#> [11,]  0.0144714388  0.122136110 0.0103188432 0.14545876 0.219901811
#> [12,]  0.0071980288  0.118620963 0.0051001437 0.12042492 0.144523861
#> [13,] -0.0002727859  0.132171139 0.0024551631 0.16977660 0.138985549
#> [14,]  0.0033361245  0.071068964 0.0039574088 0.11020989 0.123766316
#> [15,]  0.0098064179  0.084089907 0.0007633002 0.09357837 0.019882394
#> [16,]  0.0015014334 -0.006795032 0.0031348598 0.04131137 0.072743835
#> [17,]  0.0043234478  0.085748674 0.0029810821 0.13005629 0.097820041
#> [18,]  0.0086174650  0.134988586 0.0060639805 0.15775227 0.177055481
#> [19,]  0.0046697537  0.184753189 0.0155571921 0.14373525 0.147626345
#> [20,]  0.0047064935  0.098848460 0.0046434787 0.14358382 0.142492778
#> [21,]  0.0104660036  0.046091888 0.0050890638 0.03356230 0.002613545
#> [22,]  0.0135186114  0.049801037 0.0024723868 0.02410887 0.030048658
#> [23,]  0.0039882571  0.097995218 0.0031708860 0.13735858 0.105123934
#> [24,]  0.0051447619  0.118605944 0.0063457190 0.13985224 0.164558255
#> [25,]  0.0097020425  0.089817798 0.0079756360 0.10658267 0.098686493
#>                V9
#>  [1,]  0.14834504
#>  [2,]  0.26016751
#>  [3,]  0.22007686
#>  [4,]  0.15555653
#>  [5,]  0.27844881
#>  [6,]  0.25986974
#>  [7,]  0.29888940
#>  [8,]  0.12236422
#>  [9,]  0.17393814
#> [10,]  0.23488374
#> [11,]  0.27901909
#> [12,]  0.26092565
#> [13,]  0.30836360
#> [14,]  0.18465046
#> [15,]  0.05762670
#> [16,]  0.07457345
#> [17,]  0.13154951
#> [18,]  0.24192221
#> [19,]  0.02245001
#> [20,]  0.13121275
#> [21,]  0.05685182
#> [22,] -0.03773470
#> [23,]  0.09876395
#> [24,]  0.26551264
#> [25,]  0.11161584
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2028986