Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.031277763 0.32931380 0.35018405 0.32946991 0.28294770 0.15775477
#>  [2,] 0.034354700 0.20773507 0.24823746 0.29051691 0.30134905 0.27978481
#>  [3,] 0.025051882 0.36040960 0.41148162 0.41803189 0.35408041 0.22850836
#>  [4,] 0.038189348 0.24944669 0.25802102 0.28929940 0.37229210 0.37188677
#>  [5,] 0.010710205 0.23388865 0.26249483 0.34588493 0.32028884 0.24756360
#>  [6,] 0.083634354 0.24719670 0.28493510 0.37778025 0.56158984 0.69612555
#>  [7,] 0.029897334 0.28196531 0.38752306 0.50241881 0.55253862 0.57835612
#>  [8,] 0.041391810 0.08597008 0.13954710 0.15740247 0.19423433 0.27487894
#>  [9,] 0.017279331 0.25807548 0.32102991 0.30708025 0.29197371 0.25139832
#> [10,] 0.018291361 0.20430643 0.27038687 0.30720191 0.27184252 0.18770115
#> [11,] 0.016605565 0.25990384 0.29338241 0.32636409 0.36880213 0.41177898
#> [12,] 0.057091614 0.61522625 0.56087631 0.31634477 0.25407069 0.25582303
#> [13,] 0.030866618 0.08687792 0.12570901 0.12826926 0.14345292 0.23829832
#> [14,] 0.051998481 0.38657745 0.38515520 0.41937593 0.47795983 0.50199067
#> [15,] 0.019521965 0.10431833 0.09213466 0.11543733 0.22824445 0.29596333
#> [16,] 0.019420515 0.01300536 0.02986394 0.18425934 0.12358127 0.07578223
#> [17,] 0.018121170 0.10265400 0.21019117 0.39785847 0.28027590 0.24559953
#> [18,] 0.032340681 0.28946929 0.22682171 0.09863844 0.16926359 0.44462330
#> [19,] 0.008689381 0.17912527 0.14573273 0.17273600 0.28603918 0.19026039
#> [20,] 0.007935677 0.11099372 0.11030724 0.10032169 0.12483884 0.11285787
#> [21,] 0.018900000 0.25220000 0.26070000 0.37100000 0.39060000 0.26720000
#> [22,] 0.026053721 0.05149543 0.01106823 0.03789972 0.07076334 0.07810546
#> [23,] 0.027940612 0.09717235 0.15649549 0.20904195 0.18927695 0.19899901
#> [24,] 0.015640408 0.11534176 0.10185544 0.09363583 0.09010864 0.09999459
#> [25,] 0.016582378 0.06986847 0.12638541 0.08287127 0.11413546 0.19730709
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.14713933 0.16642605 0.1842808 0.2452690 0.2445087 0.040268946 0.2023531
#>  [2,] 0.24987551 0.28592599 0.2936908 0.2536087 0.3239155 0.070095234 0.3900615
#>  [3,] 0.15027316 0.20936687 0.2221888 0.2905727 0.4631745 0.023329027 0.6002279
#>  [4,] 0.33871524 0.44832262 0.5000366 0.5127111 0.6174918 0.055724454 0.7458459
#>  [5,] 0.17317494 0.20890283 0.2321010 0.3526982 0.5980722 0.007922819 0.7093188
#>  [6,] 0.73483110 0.74193045 0.7214018 0.7685930 0.8614422 0.082441074 0.8474270
#>  [7,] 0.64348967 0.75091312 0.8650847 0.9359248 0.9669650 0.038783293 0.9219924
#>  [8,] 0.39771719 0.47516289 0.6320324 0.7499154 0.8613929 0.036916338 0.9565259
#>  [9,] 0.18988840 0.15693399 0.2075252 0.3302261 0.5497389 0.060840779 0.6872779
#> [10,] 0.07821418 0.24508392 0.4144446 0.6000567 0.7257548 0.006831237 0.7344018
#> [11,] 0.44808505 0.50838251 0.6031687 0.6719827 0.7349746 0.017148004 0.8119769
#> [12,] 0.36125035 0.38086803 0.2074049 0.2158066 0.3821147 0.085049818 0.7260519
#> [13,] 0.29814123 0.42819503 0.5249541 0.6589796 0.8088379 0.021129255 0.8780603
#> [14,] 0.57897283 0.61377230 0.6861010 0.6549312 0.4716602 0.062906458 0.5868410
#> [15,] 0.29184079 0.28113631 0.1920023 0.2812510 0.6053094 0.032969738 0.8356255
#> [16,] 0.08825546 0.16215289 0.1565204 0.1028980 0.1626422 0.025056071 0.1119522
#> [17,] 0.17810192 0.17497220 0.2356437 0.4693949 0.3928934 0.070029397 0.7020716
#> [18,] 0.58511128 0.68632479 0.7306296 0.6606254 0.4658343 0.056546877 0.4039137
#> [19,] 0.06501977 0.24122995 0.3734944 0.3914787 0.5817180 0.036725813 0.6079674
#> [20,] 0.17007987 0.19817213 0.2068097 0.2463984 0.2153667 0.011963110 0.2009601
#> [21,] 0.27160000 0.41830000 0.6988000 0.5733000 0.2226000 0.030800000 0.2631000
#> [22,] 0.11645201 0.04595314 0.1153926 0.2262494 0.3294401 0.062466943 0.2854188
#> [23,] 0.26872828 0.53112744 0.7377059 0.6088687 0.3617109 0.026027224 0.3724999
#> [24,] 0.09126116 0.13408950 0.1936120 0.2773883 0.3669165 0.012399750 0.4389346
#> [25,] 0.21378614 0.20662109 0.1626384 0.1597271 0.1441408 0.011634063 0.1713579
#>             V21       V22       V23       V24        V25       V26        V27
#>  [1,] 0.3049055 0.5897424 0.7760356 0.8171051 0.84616588 0.8747803 0.92612143
#>  [2,] 0.4604272 0.4970028 0.5380066 0.6097416 0.63353477 0.6991732 0.82107948
#>  [3,] 0.7359009 0.7846687 0.7660572 0.8294788 0.89037858 0.9826555 0.98420338
#>  [4,] 0.7186907 0.7873809 0.8612266 0.8485423 0.88464350 0.9518465 0.96785116
#>  [5,] 0.8259601 0.8186958 0.8288765 0.9335367 0.93822888 0.9655950 0.99022203
#>  [6,] 0.9066878 0.8334997 0.8856820 0.9297680 0.67667762 0.5757291 0.53107837
#>  [7,] 0.7838124 0.6506730 0.5466839 0.4060177 0.29369943 0.1928303 0.17993127
#>  [8,] 0.9683523 0.9617664 0.9374617 0.7981533 0.56242864 0.4127352 0.29136884
#>  [9,] 0.8131537 0.8488983 0.8300309 0.9340476 0.95875054 0.9885659 1.01024518
#> [10,] 0.7816831 0.8351820 0.9277167 0.9310635 0.93689572 0.7761823 0.48184962
#> [11,] 0.8637444 0.9419099 0.9977050 1.0213742 0.97858988 0.9381424 0.94173164
#> [12,] 0.8144688 0.6201519 0.5393625 0.5747322 0.43195701 0.4711045 0.38043015
#> [13,] 0.8849219 0.9528962 1.0018681 0.9678554 0.92267737 0.8256779 0.73336019
#> [14,] 0.5126281 0.4428328 0.4887391 0.4922057 0.37133247 0.3569793 0.64337651
#> [15,] 0.7976813 0.7490402 0.5186456 0.4872469 0.65118666 0.7900092 0.89046727
#> [16,] 0.1943954 0.1078840 0.2054403 0.2102316 0.08839464 0.2297093 0.47413226
#> [17,] 0.8131455 0.7824766 0.7706169 0.8177920 1.00393324 0.9477986 0.70443135
#> [18,] 0.5854445 0.7648433 0.8450072 0.7925540 0.71659668 0.6660132 0.62496902
#> [19,] 0.5605370 0.4611318 0.4625316 0.4031928 0.68178959 0.9270289 0.93245868
#> [20,] 0.2688262 0.2865808 0.4249762 0.6606072 0.88258451 0.9564586 0.82419566
#> [21,] 0.7473000 0.7263000 0.3393000 0.2824000 0.60530000 0.5897000 0.49670000
#> [22,] 0.1644591 0.2212393 0.3474197 0.4442782 0.48197629 0.2574046 0.07597185
#> [23,] 0.4099246 0.4623331 0.5301676 0.7157122 0.85931333 0.8394622 0.78638149
#> [24,] 0.4691335 0.5355603 0.5943773 0.6361179 0.59947598 0.6642772 0.74031510
#> [25,] 0.2191084 0.3753054 0.5657829 0.6644868 0.71981183 0.7984708 0.79484444
#>             V28       V29         V3       V30       V31       V32       V33
#>  [1,] 0.9871815 0.8315196 0.04816787 0.5958683 0.6537150 0.6054002 0.4789775
#>  [2,] 0.8956916 0.8548493 0.06380510 0.8040511 0.7477754 0.6674159 0.5180053
#>  [3,] 0.8998511 0.7826519 0.03548349 0.6526830 0.5050043 0.3499188 0.2892436
#>  [4,] 0.9502606 0.7344449 0.03376758 0.6051614 0.5783319 0.6040112 0.4564414
#>  [5,] 0.9972159 0.8914098 0.03027921 0.6326588 0.3909360 0.2889759 0.2067090
#>  [6,] 0.4515869 0.3673576 0.06999120 0.2792218 0.1514801 0.0479650 0.2115457
#>  [7,] 0.1374109 0.1183132 0.04804196 0.2125004 0.2456658 0.3002298 0.3121782
#>  [8,] 0.3019930 0.3897086 0.04022673 0.4002931 0.3162486 0.3482460 0.3766122
#>  [9,] 0.9601572 0.8487436 0.03975022 0.6142946 0.4092932 0.3403552 0.2680407
#> [10,] 0.3863047 0.3579759 0.01524530 0.5373330 0.6961105 0.7902377 0.8814027
#> [11,] 0.9513718 0.8468491 0.04158647 0.6004496 0.3427691 0.3006869 0.1999692
#> [12,] 0.4904103 0.4814370 0.07766355 0.5681193 0.3710928 0.2862377 0.5536136
#> [13,] 0.6236558 0.4961771 0.06027895 0.4169632 0.3048616 0.1815143 0.1550355
#> [14,] 0.8619979 0.6651130 0.05721183 0.5830541 0.6149263 0.5483322 0.2707095
#> [15,] 0.7756226 0.5532965 0.02534332 0.4328819 0.3525459 0.2116422 0.2714612
#> [16,] 0.6484968 0.7089933 0.02705541 0.6886744 0.7832422 0.5655401 0.5938447
#> [17,] 0.5244264 0.4438815 0.06772746 0.4344566 0.6001990 0.6321146 0.7124440
#> [18,] 0.5313116 0.3115815 0.08038983 0.3451745 0.4608049 0.4177552 0.2849715
#> [19,] 0.5202335 0.3279865 0.05353962 0.2350076 0.1909191 0.3723409 0.4018636
#> [20,] 0.6850360 0.5548539 0.01134410 0.3870036 0.3026101 0.2214168 0.2065052
#> [21,] 0.8616000 0.8339000 0.01970000 0.4084000 0.2268000 0.1745000 0.0507000
#> [22,] 0.2632479 0.5622092 0.03779635 0.8435900 0.8448435 0.7498889 0.8105250
#> [23,] 0.8230975 0.7850884 0.03140484 0.7345374 0.8185357 0.7547659 0.6867640
#> [24,] 0.8439711 0.9372510 0.01982512 0.8664476 0.7392428 0.6403462 0.6552698
#> [25,] 0.8612841 0.8666619 0.02147650 0.7404213 0.5926245 0.4776629 0.5177404
#>              V34        V35        V36        V37        V38          V39
#>  [1,] 0.22616647 0.09863585 0.07627963 0.17943460 0.17872495  0.198075769
#>  [2,] 0.45512687 0.41677890 0.35222744 0.34075625 0.31896616  0.310579240
#>  [3,] 0.22658848 0.10582114 0.09481328 0.15206151 0.17289272  0.204162409
#>  [4,] 0.26295977 0.11107305 0.15584896 0.23302108 0.39244728  0.469471699
#>  [5,] 0.16285689 0.15789378 0.08244964 0.06619444 0.09742667  0.069882172
#>  [6,] 0.33111683 0.39010118 0.35800614 0.29001810 0.37396018  0.423071336
#>  [7,] 0.36172032 0.37101590 0.22859529 0.33253591 0.38029736  0.224629022
#>  [8,] 0.33511314 0.30271204 0.36620612 0.40589983 0.42664488  0.342756180
#>  [9,] 0.26934738 0.25739324 0.14318624 0.13001299 0.17586678  0.196524589
#> [10,] 0.83614624 0.64144808 0.43704828 0.15745102 0.20186632  0.371162471
#> [11,] 0.04009606 0.10908491 0.17292913 0.24826979 0.30824636  0.242891687
#> [12,] 0.72196335 0.75279672 0.68087669 0.64515412 0.65340982  0.599520817
#> [13,] 0.24045555 0.40211274 0.44580689 0.36165560 0.29184300  0.354923772
#> [14,] 0.15445699 0.26763912 0.40824016 0.45375558 0.37583854  0.323436539
#> [15,] 0.32846600 0.30182602 0.27391243 0.36846053 0.45521484  0.440916908
#> [16,] 0.83319020 0.95253372 0.94185022 0.74287071 0.62487857  0.642997751
#> [17,] 0.61916325 0.58662864 0.55108175 0.44280267 0.15560876 -0.005652206
#> [18,] 0.20696263 0.43384469 0.61917763 0.63376233 0.42609079  0.321755867
#> [19,] 0.29899313 0.29409813 0.38826042 0.51841294 0.39481491  0.531119438
#> [20,] 0.21550680 0.27907923 0.23551232 0.13013716 0.10358681  0.146473114
#> [21,] 0.15880000 0.30400000 0.13690000 0.16050000 0.20610000  0.073400000
#> [22,] 0.94881300 1.00338737 0.89377686 0.52639781 0.25824796  0.199507219
#> [23,] 0.75481018 0.75675626 0.72602083 0.71083545 0.58764112  0.387293389
#> [24,] 0.64960368 0.57497632 0.52951431 0.48005191 0.37577595  0.259333084
#> [25,] 0.61569994 0.56742543 0.42395330 0.18703662 0.13785000  0.173906169
#>               V4        V40        V41        V42        V43        V44
#>  [1,] 0.07185928 0.14924024 0.18891823 0.12786752 0.11914837 0.12173146
#>  [2,] 0.06278339 0.28057553 0.45510308 0.52861972 0.44000401 0.32601106
#>  [3,] 0.04775256 0.17695433 0.05884658 0.18906652 0.29637468 0.31615604
#>  [4,] 0.04913297 0.38216426 0.30918881 0.36596672 0.32510691 0.33117757
#>  [5,] 0.04555420 0.06067133 0.05521661 0.10434777 0.19380539 0.22570680
#>  [6,] 0.09575681 0.32417523 0.35660286 0.37790138 0.33785934 0.30818285
#>  [7,] 0.06993706 0.23601670 0.35283881 0.32778367 0.18394063 0.08427064
#>  [8,] 0.05453377 0.21515333 0.20094433 0.25516265 0.23045424 0.21323505
#>  [9,] 0.02408746 0.19353979 0.04962782 0.04443427 0.16579524 0.19354408
#> [10,] 0.03942552 0.52872519 0.50756315 0.37067340 0.23498789 0.11691868
#> [11,] 0.05879398 0.23282208 0.32731834 0.30921589 0.21048996 0.19685864
#> [12,] 0.08430577 0.46181418 0.32402403 0.32184091 0.27983163 0.22750839
#> [13,] 0.12087040 0.37957201 0.41177618 0.49059294 0.37490376 0.21455319
#> [14,] 0.04898856 0.29222797 0.25009200 0.22862586 0.15760390 0.10893368
#> [15,] 0.02937263 0.30133135 0.19756207 0.15930587 0.09481099 0.05881645
#> [16,] 0.01209346 0.69800579 0.58807106 0.53204600 0.41790971 0.27707904
#> [17,] 0.06658451 0.33685661 0.40970189 0.56326227 0.43499555 0.23639046
#> [18,] 0.07887834 0.42284012 0.42235137 0.32183438 0.23016190 0.18963474
#> [19,] 0.03651220 0.49052834 0.30716933 0.27011573 0.22934250 0.20542802
#> [20,] 0.02483033 0.20054106 0.24179829 0.21918712 0.17014369 0.11263525
#> [21,] 0.06220000 0.02020000 0.16380000 0.15830000 0.18300000 0.18860000
#> [22,] 0.01163054 0.22176487 0.33308809 0.39066347 0.15709979 0.22166871
#> [23,] 0.04126070 0.25491560 0.22652295 0.20435076 0.14566909 0.12046522
#> [24,] 0.03024201 0.24075164 0.21815850 0.15119157 0.14009413 0.14445598
#> [25,] 0.02958909 0.19091062 0.18368672 0.22225136 0.28248120 0.18169957
#>              V45        V46        V47        V48        V49         V5
#>  [1,] 0.08128078 0.07425109 0.10949601 0.13455572 0.07895522 0.09454083
#>  [2,] 0.39383290 0.34113257 0.23006552 0.15427849 0.09540213 0.08858949
#>  [3,] 0.26854396 0.18044953 0.14338857 0.10362577 0.05520709 0.07114588
#>  [4,] 0.34361762 0.23643692 0.15372388 0.13887276 0.07415874 0.09523255
#>  [5,] 0.21484856 0.22194945 0.17668683 0.10942035 0.05501521 0.08858848
#>  [6,] 0.21090655 0.01313809 0.05229621 0.04249314 0.03131489 0.03570457
#>  [7,] 0.09157826 0.07559053 0.05137275 0.03440481 0.02968601 0.08765439
#>  [8,] 0.13557146 0.06289104 0.04622330 0.02948447 0.03443366 0.08682746
#>  [9,] 0.14995772 0.17022057 0.12373250 0.06216171 0.03762155 0.02499601
#> [10,] 0.09665147 0.12002266 0.12167612 0.10826808 0.07687696 0.05357939
#> [11,] 0.14102158 0.10301656 0.10586928 0.09119478 0.05555813 0.11506414
#> [12,] 0.28787964 0.17840830 0.14049410 0.16925787 0.08411740 0.04313872
#> [13,] 0.13625508 0.16730297 0.16923633 0.10637477 0.05599156 0.16374377
#> [14,] 0.09250185 0.05759747 0.05985908 0.03178366 0.02403047 0.06918566
#> [15,] 0.11540690 0.12092982 0.08597074 0.04440985 0.01913367 0.04736224
#> [16,] 0.18244188 0.10022448 0.11124358 0.09150625 0.05455601 0.03981361
#> [17,] 0.10586146 0.07129918 0.06877062 0.07561961 0.03295426 0.14492004
#> [18,] 0.17357293 0.12037344 0.13110837 0.11965542 0.07194718 0.10492943
#> [19,] 0.08538134 0.05305415 0.04667635 0.01428162 0.03560181 0.04628694
#> [20,] 0.10577655 0.10520014 0.04850352 0.04273552 0.03003326 0.04587893
#> [21,] 0.10080000 0.06630000 0.01830000 0.04040000 0.01080000 0.00800000
#> [22,] 0.14234659 0.13451641 0.16350510 0.11253941 0.05642113 0.04649331
#> [23,] 0.11881714 0.14168726 0.08993322 0.04476152 0.02388626 0.05451722
#> [24,] 0.13269214 0.09465614 0.05722743 0.03881088 0.01965072 0.03722110
#> [25,] 0.05850505 0.04990757 0.04889765 0.03996990 0.02325223 0.03497251
#>               V50         V51          V52         V53         V54         V55
#>  [1,] 0.022753278 0.015832542 0.0123913403 0.007107320 0.009953811 0.004959836
#>  [2,] 0.032058815 0.029004168 0.0221202854 0.013147855 0.012123997 0.008406940
#>  [3,] 0.019809618 0.019058480 0.0125089527 0.011149710 0.008254433 0.003557090
#>  [4,] 0.022510161 0.021772453 0.0112138226 0.015557468 0.014115323 0.019520552
#>  [5,] 0.014493406 0.025275689 0.0151904965 0.005404905 0.002222674 0.003147139
#>  [6,] 0.009937763 0.027075001 0.0071814838 0.016421451 0.018048216 0.011655886
#>  [7,] 0.024056252 0.015163553 0.0273372915 0.018984943 0.015884955 0.024184447
#>  [8,] 0.015015645 0.007881171 0.0079155118 0.003625297 0.015878078 0.013415992
#>  [9,] 0.008701694 0.019490778 0.0118547418 0.006596728 0.009006729 0.004761060
#> [10,] 0.024106688 0.013471692 0.0085708647 0.004966938 0.002643818 0.006894324
#> [11,] 0.019406556 0.015091150 0.0166465199 0.014916055 0.018557111 0.006233569
#> [12,] 0.023713860 0.023375349 0.0194732433 0.015171912 0.020889311 0.015938714
#> [13,] 0.027756604 0.023996412 0.0213170323 0.023312215 0.017171711 0.003667663
#> [14,] 0.007180284 0.011587259 0.0082283236 0.005913880 0.009275206 0.010185416
#> [15,] 0.023616439 0.017923801 0.0129007760 0.021153292 0.020441166 0.010234264
#> [16,] 0.019010226 0.008750062 0.0036334027 0.015739649 0.013373417 0.008630673
#> [17,] 0.009360398 0.006923994 0.0007072671 0.012450677 0.010851886 0.013389737
#> [18,] 0.021270840 0.018887882 0.0175498174 0.013479510 0.011961248 0.007369077
#> [19,] 0.014733742 0.006515852 0.0078495131 0.004677852 0.015594027 0.018014330
#> [20,] 0.009400790 0.005767095 0.0116388982 0.005359921 0.005748789 0.003626723
#> [21,] 0.014300000 0.009100000 0.0038000000 0.009600000 0.014200000 0.019000000
#> [22,] 0.012773639 0.016229062 0.0023256212 0.004623641 0.006925306 0.006667349
#> [23,] 0.019996105 0.010990708 0.0084081785 0.004646055 0.005157111 0.012141025
#> [24,] 0.012696076 0.007687333 0.0068564172 0.008114124 0.007664194 0.003747531
#> [25,] 0.011043911 0.015920908 0.0103883274 0.011108227 0.005226493 0.004592600
#>               V56         V57         V58         V59         V6          V60
#>  [1,] 0.008786104 0.005922669 0.007099581 0.009418368 0.13805191 0.0101613206
#>  [2,] 0.008918745 0.007708968 0.010252039 0.007462881 0.11936206 0.0064150149
#>  [3,] 0.002767859 0.004229951 0.002794962 0.004332629 0.11540611 0.0045272508
#>  [4,] 0.014413897 0.012651172 0.011849456 0.010616057 0.09250856 0.0046810362
#>  [5,] 0.005485360 0.002756576 0.005226781 0.001631017 0.08564797 0.0030994395
#>  [6,] 0.015870122 0.012794781 0.011415707 0.026312331 0.13409844 0.0189714584
#>  [7,] 0.011986309 0.003564982 0.016661479 0.012273283 0.09061247 0.0075436883
#>  [8,] 0.011314874 0.011886907 0.012124388 0.011094845 0.07194939 0.0065214846
#>  [9,] 0.001672686 0.003779915 0.005586820 0.005637866 0.04646855 0.0034513951
#> [10,] 0.005127735 0.002678977 0.006897999 0.006430667 0.09235386 0.0036818445
#> [11,] 0.014042982 0.007546149 0.009699119 0.004497144 0.14469954 0.0033470516
#> [12,] 0.012126557 0.009720134 0.015821670 0.011686565 0.08554167 0.0088952614
#> [13,] 0.008899765 0.013020641 0.010624056 0.011775060 0.08959366 0.0086840026
#> [14,] 0.007967355 0.011519129 0.007263520 0.008278871 0.12668817 0.0074236129
#> [15,] 0.011003689 0.004632474 0.008562858 0.001691831 0.03040548 0.0046850005
#> [16,] 0.007006233 0.007620843 0.012358292 0.017636089 0.05716841 0.0054256392
#> [17,] 0.011792207 0.011946713 0.009200293 0.003721314 0.21989059 0.0143728619
#> [18,] 0.007122626 0.010756229 0.010079520 0.009055737 0.12792217 0.0088414100
#> [19,] 0.004268544 0.009727106 0.017238208 0.012000941 0.02975519 0.0032936887
#> [20,] 0.001739589 0.001753430 0.003364975 0.001985660 0.07713534 0.0024496580
#> [21,] 0.014000000 0.009900000 0.009200000 0.005200000 0.07890000 0.0075000000
#> [22,] 0.007086748 0.010363474 0.016160012 0.009193005 0.07912948 0.0001246429
#> [23,] 0.004788066 0.013250345 0.011093279 0.010902275 0.10792448 0.0095610240
#> [24,] 0.004412971 0.005607438 0.003801567 0.005934323 0.06729746 0.0039376274
#> [25,] 0.004661228 0.010326444 0.005873421 0.005534453 0.06302127 0.0073858897
#>               V7           V8          V9
#>  [1,] 0.15278534 0.1933035864  0.24279571
#>  [2,] 0.11288641 0.1486855655  0.22087255
#>  [3,] 0.12505809 0.1568784628  0.25951479
#>  [4,] 0.11616984 0.1372158490  0.18966558
#>  [5,] 0.09869795 0.1279119069  0.19412013
#>  [6,] 0.22832288 0.1942105317  0.12052986
#>  [7,] 0.11162767 0.0910820646  0.14476751
#>  [8,] 0.09191977 0.0900647343  0.13927805
#>  [9,] 0.10007518 0.1078505408  0.18318426
#> [10,] 0.09710776 0.1645827456  0.22405330
#> [11,] 0.13519552 0.1462257678  0.22890633
#> [12,] 0.17706405 0.3235580857  0.50274601
#> [13,] 0.11068397 0.1323912879  0.05057618
#> [14,] 0.16774036 0.2500902632  0.35169515
#> [15,] 0.04458086 0.0598794253  0.11035744
#> [16,] 0.04506636 0.0554665832 -0.01347698
#> [17,] 0.15197156 0.1549977214  0.04150439
#> [18,] 0.13212061 0.1754070318  0.21993842
#> [19,] 0.04074086 0.0243269715  0.08174781
#> [20,] 0.08730587 0.0730800968  0.10428650
#> [21,] 0.14400000 0.1451000000  0.17890000
#> [22,] 0.08459904 0.0008457745  0.03658064
#> [23,] 0.13730154 0.1356573827  0.06630270
#> [24,] 0.06308154 0.0739221856  0.10048773
#> [25,] 0.05023674 0.0684881769  0.08161308
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551