Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11         V12         V13        V14
#>  [1,] 0.032384942 0.22889113 0.27308014  0.28843788 0.213350115 0.11918537
#>  [2,] 0.026713776 0.39218928 0.38933501  0.32874918 0.288876099 0.17161827
#>  [3,] 0.021824283 0.37154749 0.46783222  0.50879560 0.533627194 0.40612143
#>  [4,] 0.056709339 0.12084511 0.20907393  0.24450527 0.279293600 0.38761209
#>  [5,] 0.092664659 0.56610173 0.46643878  0.19709877 0.229216853 0.29545372
#>  [6,] 0.039759619 0.22620920 0.22006040  0.28928542 0.323398858 0.25798486
#>  [7,] 0.026135610 0.24723694 0.24842843  0.28535501 0.346651688 0.39795618
#>  [8,] 0.070700000 0.12470000 0.23400000  0.17640000 0.228400000 0.31150000
#>  [9,] 0.023427356 0.57573330 0.65049917  0.53552853 0.456406852 0.29618992
#> [10,] 0.025489326 0.25730430 0.30569550  0.35426738 0.309723163 0.20423285
#> [11,] 0.036724098 0.22373687 0.30029652  0.31913454 0.274278404 0.13088494
#> [12,] 0.017192169 0.30610390 0.33970152  0.30945648 0.306316143 0.26212993
#> [13,] 0.011054594 0.05983965 0.08770662  0.07870454 0.204777191 0.26698294
#> [14,] 0.008446860 0.11095864 0.08906701  0.07444209 0.085118612 0.08384755
#> [15,] 0.011358287 0.09933039 0.10328203  0.06063689 0.157939115 0.34248293
#> [16,] 0.024062667 0.10027764 0.13197403  0.15723991 0.233372139 0.38109576
#> [17,] 0.008231586 0.11132449 0.10210799 -0.01940606 0.003041512 0.06155695
#> [18,] 0.034472616 0.29486671 0.25160620  0.10534958 0.210328400 0.48483435
#> [19,] 0.024415839 0.06293413 0.06260005  0.10931930 0.110026231 0.11117042
#> [20,] 0.019452226 0.10692561 0.14689709  0.11854461 0.131257456 0.11791071
#> [21,] 0.036741496 0.35394476 0.38149338  0.44597897 0.583911302 0.66256672
#> [22,] 0.024253096 0.12002782 0.12743417  0.06418352 0.060245067 0.12816248
#> [23,] 0.022291621 0.12418340 0.13351240  0.18578500 0.168684900 0.18213794
#> [24,] 0.014399254 0.08991034 0.09244877  0.08502270 0.063269957 0.14026593
#> [25,] 0.004785122 0.08323374 0.08495348  0.09620301 0.145880409 0.14428729
#>              V15        V16        V17       V18       V19            V2
#>  [1,] 0.15461896 0.18295807 0.21083278 0.2416661 0.3159541  0.0496558798
#>  [2,] 0.17425466 0.20519808 0.21248386 0.2688525 0.2411119  0.0376169511
#>  [3,] 0.32009101 0.32761274 0.31485740 0.2827771 0.4813087  0.0482255338
#>  [4,] 0.53674904 0.64185681 0.80278846 0.7536295 0.7739491  0.0620656580
#>  [5,] 0.56387565 0.58201565 0.17198900 0.1090929 0.3578786  0.0984315245
#>  [6,] 0.22737287 0.30957785 0.27948148 0.1763000 0.3027555  0.0754982062
#>  [7,] 0.45220016 0.58114935 0.68665930 0.7510502 0.7472503  0.0469302970
#>  [8,] 0.47250000 0.55430000 0.53860000 0.3746000 0.4583000  0.1252000000
#>  [9,] 0.18021684 0.23683111 0.28424403 0.3774692 0.5366877  0.0328537543
#> [10,] 0.16211137 0.25760372 0.29072533 0.4107557 0.5602755  0.0179838186
#> [11,] 0.11122813 0.15271082 0.18614277 0.2408992 0.2084874  0.0175698346
#> [12,] 0.16140441 0.11612962 0.26453046 0.4354086 0.6533562  0.0442842940
#> [13,] 0.26195897 0.26700694 0.27167128 0.3913516 0.5276605  0.0104472893
#> [14,] 0.13540685 0.14443365 0.10613977 0.1276388 0.1163177  0.0127205948
#> [15,] 0.48154320 0.65958051 0.77947396 0.7901931 0.8217682  0.0106077611
#> [16,] 0.53952204 0.70543570 0.82135871 0.8462310 0.8493624  0.0252996037
#> [17,] 0.17046450 0.08204331 0.07782425 0.1611076 0.2660011  0.0190060521
#> [18,] 0.61596365 0.73598504 0.78239184 0.6956161 0.5131870  0.0832203843
#> [19,] 0.10879586 0.10711640 0.12374568 0.1543229 0.2308953  0.0499652978
#> [20,] 0.06328553 0.08326713 0.11802963 0.1819406 0.4027621  0.0139133563
#> [21,] 0.75447160 0.75616799 0.60279357 0.5013499 0.4560020  0.0700710230
#> [22,] 0.07796313 0.10212753 0.21785497 0.2237764 0.2473095  0.0084991063
#> [23,] 0.22545448 0.35685493 0.47317561 0.4385477 0.3638182  0.0311789420
#> [24,] 0.07433338 0.04838512 0.17334479 0.2860826 0.2954234 -0.0008343641
#> [25,] 0.17320933 0.15890739 0.13798133 0.1550776 0.1103174  0.0058756902
#>             V20       V21       V22       V23       V24       V25       V26
#>  [1,] 0.3084088 0.4537812 0.6206809 0.6910966 0.7401433 0.7767245 0.8423083
#>  [2,] 0.1833337 0.2877965 0.6212442 0.8021006 0.8321008 0.8467400 0.8540105
#>  [3,] 0.7448004 0.8297945 0.8056072 0.6715529 0.6761937 0.7269094 0.8744250
#>  [4,] 0.8302583 0.8927245 0.9039597 0.7708324 0.5837699 0.4410516 0.4295516
#>  [5,] 0.8743745 0.8716008 0.3822428 0.1573259 0.1701205 0.2041516 0.3366565
#>  [6,] 0.3528278 0.3968680 0.4342915 0.5102319 0.5784046 0.5564987 0.5786841
#>  [7,] 0.8039304 0.8385212 0.9091581 0.9969061 1.0039311 0.9605567 0.9234693
#>  [8,] 0.5961000 0.7464000 0.7644000 0.5711000 0.6257000 0.6695000 0.7131000
#>  [9,] 0.6656229 0.8225670 0.8750580 0.8392319 0.8503882 0.8012596 0.8103068
#> [10,] 0.6589277 0.7653382 0.7476052 0.8052599 0.9035563 0.9068100 0.9273740
#> [11,] 0.2172683 0.3247990 0.5935439 0.8152826 0.8917835 0.9784845 0.9928451
#> [12,] 0.7686278 0.6944017 0.4462067 0.2290693 0.2643930 0.5041073 0.8483824
#> [13,] 0.6095836 0.6026123 0.5598267 0.5352257 0.6543300 0.8301809 0.8440343
#> [14,] 0.1406430 0.2346778 0.3295178 0.4729737 0.6614463 0.8667460 0.9690150
#> [15,] 0.8155606 0.8252683 0.8491369 0.8821412 0.8884960 0.9008551 0.9660089
#> [16,] 0.8468416 0.9492079 0.9484520 0.8479771 0.8259149 0.8496839 0.8661688
#> [17,] 0.2905473 0.3779944 0.7935471 0.9119330 0.8427997 0.6634665 0.6960461
#> [18,] 0.4442298 0.6045658 0.7511371 0.8059502 0.8234024 0.7564947 0.6662747
#> [19,] 0.2450755 0.1836229 0.1603801 0.2552876 0.2720286 0.2540355 0.3146680
#> [20,] 0.3829349 0.2438411 0.2647184 0.4181853 0.5840436 0.5540843 0.5278562
#> [21,] 0.5581515 0.5252336 0.4332107 0.3271239 0.4793309 0.5067619 0.5441617
#> [22,] 0.3827321 0.5311460 0.6836382 0.8620777 0.9009904 0.8820163 0.9442887
#> [23,] 0.4030919 0.4226411 0.4419797 0.4782836 0.5944487 0.6189983 0.5895453
#> [24,] 0.5041721 0.6268356 0.7551503 0.9106542 0.8930133 0.9407750 0.9854756
#> [25,] 0.1269352 0.2035895 0.2060557 0.3438126 0.5881060 0.8632468 1.0108567
#>             V27       V28       V29          V3       V30       V31       V32
#>  [1,] 0.9128433 0.8627783 0.7676829 0.054552987 0.7122083 0.6763795 0.5264250
#>  [2,] 0.9214849 1.0000000 0.7907898 0.046969511 0.5418364 0.6447346 0.6450200
#>  [3,] 0.9936624 0.9680041 0.8695301 0.048951401 0.6102230 0.5239219 0.5965572
#>  [4,] 0.3844809 0.2848474 0.2795443 0.060821351 0.3828311 0.2241972 0.3078888
#>  [5,] 0.2574657 0.5642377 0.5252494 0.050934741 0.6841891 0.3489053 0.4040511
#>  [6,] 0.7227648 0.8473473 0.8143034 0.052799287 0.8565420 0.8352293 0.6964122
#>  [7,] 0.8963808 0.8313134 0.7186593 0.049777425 0.5723863 0.3731179 0.3013792
#>  [8,] 0.7567000 0.8077000 0.8477000 0.144700000 0.9289000 0.9513000 0.7995000
#>  [9,] 0.8113039 0.8221860 0.6634640 0.042560307 0.4476882 0.3546590 0.3484968
#> [10,] 0.9532812 0.9888913 0.8801502 0.033753106 0.6447680 0.3909095 0.2635652
#> [11,] 0.9503801 1.0000000 0.8989290 0.027176665 0.7315737 0.7696780 0.7158796
#> [12,] 1.0287079 0.9332529 0.8686651 0.046443871 0.7964830 0.5865587 0.5160388
#> [13,] 0.9096158 0.9364753 0.7721817 0.002096073 0.6290807 0.5839862 0.4549455
#> [14,] 0.8196602 0.6053222 0.4911133 0.004900864 0.3544236 0.2643081 0.1367542
#> [15,] 1.0056114 0.9154956 0.6770126 0.010812953 0.4722548 0.2954346 0.2901852
#> [16,] 0.7697883 0.5715185 0.4628407 0.029686476 0.5050424 0.4842652 0.3428037
#> [17,] 0.5951437 0.6938506 0.8150587 0.041635519 0.7660251 0.5656911 0.5538379
#> [18,] 0.6351846 0.5149694 0.3426022 0.094736083 0.3503869 0.4089078 0.3326479
#> [19,] 0.3639610 0.5171373 0.6478980 0.029708329 0.7605557 0.7877888 0.6706186
#> [20,] 0.4452313 0.6724775 0.9898042 0.017359504 1.0072696 0.9537162 0.8797497
#> [21,] 0.6126034 0.7692433 0.7017294 0.058657397 0.6977071 0.7072579 0.5025511
#> [22,] 0.9503553 0.8839804 0.8843862 0.012438930 0.7767312 0.7206362 0.6621519
#> [23,] 0.6357350 0.7377967 0.8057342 0.029128416 0.7911883 0.7544024 0.6649158
#> [24,] 0.9433329 0.8514235 0.7863416 0.018228102 0.6846912 0.6343754 0.5848979
#> [25,] 0.8661340 0.7147050 0.6043454 0.008879497 0.4439791 0.3766850 0.2403288
#>             V33       V34        V35        V36        V37        V38
#>  [1,] 0.3533871 0.2537134 0.15660696 0.15929864 0.26897309 0.21953363
#>  [2,] 0.4913295 0.2134786 0.09819344 0.04439533 0.16657271 0.19465088
#>  [3,] 0.5002615 0.2956193 0.06011512 0.27473795 0.25187098 0.39662159
#>  [4,] 0.3653758 0.2711935 0.26478371 0.23758906 0.34067924 0.38009366
#>  [5,] 0.6785845 0.8011721 0.90932957 0.88395474 0.87943774 0.90514036
#>  [6,] 0.4563799 0.4632876 0.48715171 0.41044547 0.39148981 0.39466614
#>  [7,] 0.2072677 0.1555733 0.20662371 0.20951614 0.25474392 0.26047813
#>  [8,] 0.4362000 0.4048000 0.49520000 0.17120000 0.36520000 0.37630000
#>  [9,] 0.4177104 0.4404511 0.30403681 0.38420841 0.31943176 0.33775027
#> [10,] 0.1913778 0.1123498 0.09508671 0.02903397 0.06754011 0.08960982
#> [11,] 0.5280212 0.2054301 0.09657938 0.13301252 0.11383246 0.12735757
#> [12,] 0.7018533 0.6480600 0.51281420 0.34068822 0.36546009 0.38020823
#> [13,] 0.3863882 0.2457731 0.17976629 0.29109163 0.27741282 0.33100732
#> [14,] 0.1735062 0.2099150 0.22684037 0.18081118 0.08595146 0.08623382
#> [15,] 0.1676287 0.2140795 0.26562299 0.23565604 0.21902442 0.16659654
#> [16,] 0.3211047 0.2949558 0.16437474 0.08479386 0.14742429 0.13940778
#> [17,] 0.7450677 0.9145777 0.89388609 0.68583053 0.40532581 0.23194993
#> [18,] 0.2776818 0.2037070 0.44114717 0.67000590 0.69121154 0.43255844
#> [19,] 0.7290588 0.8222171 0.87939708 0.85038838 0.66657871 0.54320279
#> [20,] 0.7653082 0.6539356 0.53548759 0.44105832 0.44991311 0.16728863
#> [21,] 0.1814254 0.2279048 0.38636660 0.44145593 0.32534543 0.32606776
#> [22,] 0.6087897 0.6399224 0.63516862 0.53243141 0.44176914 0.35488682
#> [23,] 0.6518022 0.7205584 0.71702151 0.68830426 0.63523992 0.50311173
#> [24,] 0.5677714 0.5540066 0.52821134 0.48469326 0.43159242 0.29691488
#> [25,] 0.2093854 0.2534593 0.35276042 0.25876722 0.08573364 0.04857396
#>              V39          V4        V40        V41        V42        V43
#>  [1,] 0.18770133 0.054046691 0.25603923 0.28667951 0.27013735 0.25037064
#>  [2,] 0.18683306 0.080961694 0.11132627 0.16733414 0.06107985 0.07145945
#>  [3,] 0.60140329 0.040049006 0.49008916 0.25535991 0.43780550 0.41581284
#>  [4,] 0.30057758 0.065191051 0.17858833 0.14990049 0.30240265 0.25609801
#>  [5,] 0.71811824 0.088793324 0.48312246 0.27923216 0.24278878 0.20040053
#>  [6,] 0.39609146 0.040440922 0.40364602 0.57843064 0.57159343 0.51107526
#>  [7,] 0.23465193 0.052105579 0.26491338 0.34281285 0.30212584 0.16838812
#>  [8,] 0.28410000 0.164400000 0.04270000 0.53310000 0.69520000 0.42880000
#>  [9,] 0.50616297 0.062594597 0.42487240 0.19746721 0.28716394 0.39058057
#> [10,] 0.08845787 0.054910170 0.11835144 0.09845103 0.11427870 0.19490941
#> [11,] 0.18037930 0.050352991 0.03405537 0.15459932 0.07137310 0.10129423
#> [12,] 0.33957863 0.070090296 0.28254441 0.38310441 0.39435943 0.36088981
#> [13,] 0.34398787 0.014131974 0.19362532 0.08094794 0.19942582 0.15896887
#> [14,] 0.18667226 0.007879838 0.24007538 0.29580271 0.29077936 0.21229494
#> [15,] 0.04284101 0.036560730 0.01441822 0.04258276 0.01623141 0.05604161
#> [16,] 0.13549392 0.042924054 0.20476821 0.13705117 0.08676143 0.05003356
#> [17,] 0.22752840 0.039840419 0.30279362 0.26842943 0.20287934 0.32790729
#> [18,] 0.26006588 0.097632128 0.38439973 0.42257021 0.36290575 0.23847750
#> [19,] 0.49742916 0.005238652 0.45916592 0.38178531 0.38396319 0.24704341
#> [20,] 0.02949121 0.041985444 0.11163821 0.13843153 0.14252424 0.24159907
#> [21,] 0.31971732 0.037982628 0.34956494 0.28974415 0.15513010 0.15669754
#> [22,] 0.22424927 0.018692196 0.20951838 0.18483674 0.07983332 0.08213060
#> [23,] 0.34056404 0.035519361 0.27552086 0.22608672 0.19520001 0.15728303
#> [24,] 0.07621754 0.044589862 0.16471023 0.19193902 0.12346758 0.06049084
#> [25,] 0.08255417 0.029805013 0.15640076 0.22998261 0.22592723 0.19971587
#>                V44        V45          V46           V47         V48
#>  [1,]  0.143857178 0.14574550  0.149292173  0.1641005186 0.143937770
#>  [2,]  0.138802546 0.06531015  0.060983020  0.0691338445 0.106224769
#>  [3,]  0.458674070 0.50225837  0.202188424  0.1216102291 0.163756556
#>  [4,]  0.171293495 0.13056323  0.066440043  0.0584767313 0.057353628
#>  [5,]  0.272583752 0.42340516  0.303315239  0.1488986855 0.099178579
#>  [6,]  0.418331850 0.52203781  0.511972822  0.3537479977 0.220986650
#>  [7,]  0.188081619 0.16069058  0.116160678  0.1006399070 0.074237567
#>  [8,]  0.306300000 0.58350000  0.569200000  0.2630000000 0.119600000
#>  [9,]  0.386287003 0.30015506  0.121188644  0.1254749747 0.153444831
#> [10,]  0.227913088 0.19342153  0.198816935  0.1663975637 0.102846274
#> [11,]  0.160840815 0.09442181  0.023050827  0.0849803140 0.119942087
#> [12,]  0.212978787 0.10203077  0.078633840  0.0689897607 0.048964240
#> [13,]  0.125150260 0.12321907  0.097242583  0.1057179429 0.077297671
#> [14,]  0.138656323 0.10942965  0.118333527  0.0607715346 0.032791385
#> [15,]  0.038405573 0.03109688 -0.006984205  0.0298419897 0.026433845
#> [16,]  0.107745920 0.11219159  0.078733029  0.1068488278 0.079350032
#> [17,]  0.226368646 0.08205892  0.023444535  0.0323032295 0.035231606
#> [18,]  0.205692476 0.19102133  0.139039510  0.1376361606 0.132651944
#> [19,]  0.220492679 0.17777290  0.115344553  0.1253563033 0.095226220
#> [20,]  0.369344187 0.36537543  0.228437337  0.0914090854 0.006411654
#> [21,]  0.134069051 0.09844726  0.055428681  0.0628551769 0.028827920
#> [22,]  0.050384817 0.01512987  0.025326159  0.0253477746 0.036336786
#> [23,]  0.165263840 0.13790428  0.127686307  0.0853549653 0.052128793
#> [24,] -0.002997177 0.02443241 -0.004161446 -0.0001846602 0.020033383
#> [25,]  0.102111612 0.07959160  0.082835667  0.0328565071 0.038943766
#>               V49         V5         V50         V51         V52         V53
#>  [1,] 0.081804237 0.05770491 0.018364798 0.020132945 0.019501906 0.009586259
#>  [2,] 0.073920755 0.10850132 0.027585804 0.014926056 0.009243007 0.004803175
#>  [3,] 0.093695977 0.05949924 0.016707584 0.032454205 0.017229669 0.003637405
#>  [4,] 0.035824097 0.10717192 0.017826623 0.011867478 0.011398226 0.010602891
#>  [5,] 0.056240502 0.07302441 0.025222110 0.021734999 0.019648847 0.012428488
#>  [6,] 0.134057353 0.06768487 0.054325489 0.040402816 0.026720305 0.013010385
#>  [7,] 0.041734150 0.09937504 0.019133394 0.017072449 0.014566035 0.017298273
#>  [8,] 0.098300000 0.16930000 0.037400000 0.029100000 0.015600000 0.019700000
#>  [9,] 0.085383977 0.08150635 0.025477124 0.024192317 0.013951463 0.015640961
#> [10,] 0.048698918 0.06446902 0.013408839 0.016873667 0.012767863 0.005024344
#> [11,] 0.062210055 0.06690725 0.019144972 0.007412092 0.008300933 0.007877174
#> [12,] 0.021743492 0.17189824 0.011536353 0.012153658 0.012610550 0.011984895
#> [13,] 0.037610050 0.02577915 0.024151726 0.006452683 0.008043057 0.006881274
#> [14,] 0.028011327 0.02384086 0.009731232 0.003645793 0.006196905 0.002635161
#> [15,] 0.017999712 0.06554307 0.011500375 0.012870635 0.007788770 0.002259881
#> [16,] 0.035870532 0.08234672 0.017917858 0.010120305 0.010677218 0.009126384
#> [17,] 0.007832158 0.07256018 0.014326314 0.018296504 0.014876355 0.020746034
#> [18,] 0.082279352 0.11173730 0.021808674 0.018939542 0.017160893 0.011170702
#> [19,] 0.051197291 0.02591683 0.013352644 0.010147206 0.008821478 0.014274005
#> [20,] 0.008095039 0.03644855 0.016892552 0.011480823 0.001979881 0.006843322
#> [21,] 0.014062801 0.12774083 0.010384923 0.009514577 0.013427076 0.007023259
#> [22,] 0.008825153 0.05509916 0.007381153 0.004195795 0.003220978 0.009150438
#> [23,] 0.023586341 0.06590110 0.020351173 0.013795937 0.010725876 0.008470748
#> [24,] 0.003289541 0.04739644 0.007706383 0.005997308 0.007376609 0.005700289
#> [25,] 0.031274966 0.04397080 0.006007350 0.008766130 0.012037425 0.005367555
#>               V54          V55          V56         V57         V58
#>  [1,] 0.013666350 0.0058087014 0.0081842371 0.006778814 0.009218391
#>  [2,] 0.005192363 0.0042427972 0.0098171608 0.005462923 0.005466098
#>  [3,] 0.009564472 0.0124611662 0.0035187009 0.006658722 0.004246986
#>  [4,] 0.014596356 0.0086813321 0.0133341757 0.005232465 0.007085781
#>  [5,] 0.024707486 0.0229738812 0.0117110732 0.010932412 0.015020821
#>  [6,] 0.014342830 0.0120357587 0.0117263704 0.007849083 0.012230745
#>  [7,] 0.015766385 0.0094391963 0.0111163539 0.009311951 0.008240297
#>  [8,] 0.013500000 0.0127000000 0.0138000000 0.013300000 0.013100000
#>  [9,] 0.016623917 0.0096587738 0.0068708883 0.004830087 0.008253974
#> [10,] 0.004585465 0.0044015108 0.0044372650 0.003706134 0.002687719
#> [11,] 0.011632541 0.0062777683 0.0068780229 0.005756555 0.005556555
#> [12,] 0.009480797 0.0016628217 0.0017481369 0.001081021 0.004155929
#> [13,] 0.010730704 0.0088817363 0.0068713764 0.006942406 0.005638115
#> [14,] 0.003882210 0.0040312007 0.0032256673 0.001500286 0.002639887
#> [15,] 0.012301489 0.0059526999 0.0078986304 0.007724178 0.005450214
#> [16,] 0.006149234 0.0060949606 0.0111930092 0.008138312 0.002310773
#> [17,] 0.016112975 0.0056837174 0.0143837987 0.022726471 0.008121654
#> [18,] 0.014690369 0.0097345526 0.0061825868 0.009230470 0.009602517
#> [19,] 0.010837600 0.0118628854 0.0069022580 0.006908173 0.010813470
#> [20,] 0.012127397 0.0004833909 0.0016426373 0.005600942 0.004375171
#> [21,] 0.007362545 0.0060254660 0.0125990285 0.008360305 0.004617251
#> [22,] 0.008757396 0.0064368394 0.0072802353 0.005881308 0.005337422
#> [23,] 0.011111684 0.0116613696 0.0062806259 0.010959450 0.006811198
#> [24,] 0.001335272 0.0019930085 0.0073054632 0.004590848 0.001276633
#> [25,] 0.002814514 0.0025458728 0.0001284083 0.002014026 0.003534204
#>                V59         V6           V60         V7         V8         V9
#>  [1,] 0.0098242699 0.13464581  0.0097765781 0.13493469 0.11607994 0.17394406
#>  [2,] 0.0072321956 0.11331459  0.0106088950 0.13878407 0.23002186 0.30680958
#>  [3,] 0.0036642016 0.10016008  0.0041927420 0.11315446 0.17964222 0.26105391
#>  [4,] 0.0097577616 0.08237953  0.0089838426 0.07096767 0.04852838 0.11356726
#>  [5,] 0.0136137175 0.12135571  0.0131387615 0.26621503 0.44909865 0.58401509
#>  [6,] 0.0087825977 0.12539348  0.0067109173 0.13806662 0.18693774 0.23233324
#>  [7,] 0.0062116544 0.13217818  0.0054732489 0.14006234 0.16171514 0.22612201
#>  [8,] 0.0154000000 0.08440000  0.0218000000 0.07150000 0.09470000 0.15830000
#>  [9,] 0.0087295046 0.13653022  0.0058094546 0.17886615 0.25156955 0.39572143
#> [10,] 0.0009270704 0.10685989  0.0032902030 0.15432121 0.16382518 0.19873822
#> [11,] 0.0065334748 0.10053534  0.0058996606 0.10493585 0.10535706 0.10806882
#> [12,] 0.0074206261 0.09249369  0.0020849207 0.05530548 0.10626987 0.32435367
#> [13,] 0.0010515318 0.01072722  0.0023050222 0.06304894 0.05773219 0.07896375
#> [14,] 0.0006314709 0.02623610  0.0030940454 0.04939089 0.04828503 0.09744628
#> [15,] 0.0058802014 0.08636705  0.0043452418 0.10206254 0.06105306 0.08479781
#> [16,] 0.0058938842 0.10222466  0.0051085648 0.12883971 0.10492271 0.11089425
#> [17,] 0.0077871827 0.07036744  0.0137082788 0.10490262 0.08222537 0.02114665
#> [18,] 0.0094856540 0.12845146  0.0071726257 0.15578585 0.18767202 0.23024300
#> [19,] 0.0074276814 0.05766899 -0.0002729892 0.06496370 0.05682275 0.05093035
#> [20,] 0.0068512842 0.02025000  0.0010089042 0.08649281 0.09471224 0.07408691
#> [21,] 0.0073929701 0.28120224  0.0061257591 0.28404033 0.33089926 0.38288838
#> [22,] 0.0021989184 0.09619316  0.0027229805 0.09003739 0.08788248 0.14114310
#> [23,] 0.0088840919 0.11675579  0.0060367566 0.13886276 0.12193346 0.12596005
#> [24,] 0.0034646177 0.08131745  0.0045842478 0.09802479 0.10970637 0.14086803
#> [25,] 0.0013966366 0.10545043  0.0016753912 0.09928579 0.07744566 0.11111131
#> 
#> $cl
#>  [1] M M M M M M M M M M M M R R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1594203