Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                  V1        V10        V11          V12        V13        V14
#>  [1,]  3.670217e-02 0.36001391 0.41976080  0.394642353 0.43673745 0.51914322
#>  [2,]  3.643564e-02 0.31729908 0.36553057  0.382982120 0.30597358 0.20603424
#>  [3,]  1.830269e-02 0.14237028 0.04057884 -0.027255913 0.05693145 0.18828658
#>  [4,]  3.222928e-02 0.30073876 0.36929124  0.414690677 0.43025135 0.37843717
#>  [5,]  1.659627e-02 0.32121092 0.27979403  0.292443262 0.24106214 0.17093552
#>  [6,]  8.093932e-02 0.11271097 0.18022972  0.346880551 0.36139520 0.23382659
#>  [7,]  4.243192e-02 0.22746709 0.20625783  0.270529022 0.31461858 0.30194020
#>  [8,] -6.688345e-05 0.23389696 0.05967910 -0.063633527 0.16651598 0.29909997
#>  [9,]  2.000489e-02 0.10738007 0.16356464  0.137249130 0.12662005 0.18982068
#> [10,]  2.141292e-02 0.19701206 0.29535725  0.293915819 0.24159228 0.28208748
#> [11,]  6.632106e-02 0.21301316 0.36047226  0.509423734 0.64023297 0.73004439
#> [12,]  5.239571e-02 0.21860858 0.32456493  0.355902108 0.40320731 0.41797001
#> [13,]  8.775252e-02 0.56202817 0.45612734  0.191562501 0.22377046 0.30272074
#> [14,]  1.933559e-02 0.17309111 0.16196226  0.200677100 0.31540621 0.24387141
#> [15,]  1.837725e-02 0.07553577 0.11700298  0.237766631 0.20320327 0.20771994
#> [16,]  2.790473e-02 0.29890518 0.27063624  0.083548950 0.24099856 0.50217857
#> [17,]  5.385901e-02 0.23007523 0.28368543  0.355879306 0.31671507 0.19564833
#> [18,]  2.930000e-02 0.08600000 0.04140000  0.047200000 0.08350000 0.09380000
#> [19,]  1.462538e-02 0.13842278 0.08081313  0.009777738 0.00738409 0.07848679
#> [20,]  1.803966e-02 0.07091252 0.06086793  0.150819360 0.12102569 0.11120950
#> [21,]  7.676110e-03 0.12020372 0.15807699  0.129863998 0.15919597 0.16850959
#> [22,]  3.495198e-02 0.34430461 0.28521062  0.078917913 0.17251767 0.49916651
#> [23,]  2.719777e-02 0.15220483 0.16700417  0.135560530 0.13503911 0.13436015
#> [24,]  1.578621e-02 0.09840174 0.11648188  0.125906363 0.14121148 0.13477951
#> [25,]  2.483010e-02 0.11004259 0.14363100  0.058964026 0.13502654 0.15362523
#>              V15       V16       V17       V18        V19           V2
#>  [1,] 0.46524321 0.3889582 0.4319343 0.5292779 0.65288119  0.041626972
#>  [2,] 0.17604939 0.2530685 0.2690152 0.4006809 0.59703915  0.044869139
#>  [3,] 0.38357390 0.5933690 0.7064067 0.7289777 0.83229889  0.008924407
#>  [4,] 0.25229924 0.2833485 0.3154008 0.3395530 0.43635680  0.044102539
#>  [5,] 0.15884619 0.1991410 0.1915446 0.2093393 0.16883937  0.020496932
#>  [6,] 0.20994616 0.2736786 0.1892667 0.2488078 0.36908361  0.138499625
#>  [7,] 0.37008555 0.4759086 0.4322554 0.2096185 0.42953675  0.097320661
#>  [8,] 0.42651527 0.6089617 0.5982090 0.6697048 0.83957346 -0.003416593
#>  [9,] 0.21360589 0.3915598 0.4975641 0.6069164 0.75372845 -0.001895628
#> [10,] 0.23389505 0.1791993 0.1783551 0.2026994 0.33763477  0.027664880
#> [11,] 0.83446725 0.8737933 0.9626109 0.9691213 0.92192987  0.069518986
#> [12,] 0.52843531 0.5659906 0.6545078 0.7229728 0.84576745  0.059147066
#> [13,] 0.54081944 0.5567822 0.1664235 0.1097547 0.34702195  0.097942568
#> [14,] 0.20638909 0.3458722 0.4389259 0.4919356 0.67962058  0.049044749
#> [15,] 0.21001075 0.3818093 0.7009467 0.6287599 0.34079757  0.030802795
#> [16,] 0.63032818 0.7693224 0.8618780 0.7544423 0.52647315  0.093229800
#> [17,] 0.26154036 0.3695138 0.5061737 0.6105524 0.62987514  0.054780644
#> [18,] 0.14660000 0.0809000 0.1179000 0.2179000 0.33260000  0.064400000
#> [19,] 0.10632489 0.1596680 0.2366978 0.2439068 0.31321742  0.001953320
#> [20,] 0.10940432 0.1361233 0.1213907 0.1614224 0.15491128  0.025852694
#> [21,] 0.23055433 0.2741527 0.2150987 0.1782934 0.19982323  0.008760615
#> [22,] 0.64834802 0.6834153 0.7503363 0.6479194 0.43916374  0.031192272
#> [23,] 0.21638051 0.5407977 0.6883369 0.5620963 0.30594783  0.009160446
#> [24,] 0.07932786 0.1081023 0.1486033 0.1992170 0.32120571  0.011892582
#> [25,] 0.18304151 0.1918934 0.1535304 0.1325519 0.09501674  0.018530952
#>             V20       V21       V22       V23       V24        V25       V26
#>  [1,] 0.7166423 0.7735133 0.8798783 0.9418656 0.9541940 0.87444278 0.9202067
#>  [2,] 0.7224091 0.8172864 0.8051784 0.8694516 0.9554154 0.91788619 0.9579770
#>  [3,] 0.9585362 0.9469777 0.9814575 0.8542316 0.5285326 0.28822321 0.2270923
#>  [4,] 0.6602223 0.7189363 0.7278927 0.7093469 0.7047782 0.69253836 0.8008129
#>  [5,] 0.1215294 0.1118790 0.2520782 0.4485739 0.6051045 0.72442784 0.8364033
#>  [6,] 0.1619711 0.2446022 0.6693033 1.0032079 0.9889421 0.79506842 0.6389903
#>  [7,] 0.4474948 0.5661910 0.5070910 0.4581833 0.5897046 0.55864655 0.5558372
#>  [8,] 0.9136557 1.0360528 1.0703661 1.0310097 0.6957232 0.39089325 0.4258155
#>  [9,] 0.8285788 0.8623849 0.9250208 0.9935791 0.9890956 0.99978542 0.9463681
#> [10,] 0.3250761 0.3919925 0.4100936 0.4217786 0.5285183 0.57125530 0.7041774
#> [11,] 0.8031842 0.6792757 0.5954639 0.4808063 0.3331061 0.25217018 0.1926110
#> [12,] 0.8548844 0.9558925 0.9626505 0.9131305 0.8580374 0.68092943 0.5399307
#> [13,] 0.8796085 0.8856781 0.3866893 0.1577314 0.1707932 0.20154409 0.3564078
#> [14,] 0.7411793 0.7131650 0.5021168 0.4685650 0.5269893 0.68559068 0.7790394
#> [15,] 0.2346805 0.3558403 0.5242898 0.6307673 0.7083136 0.82116552 0.8911149
#> [16,] 0.3892207 0.5326843 0.7107926 0.8095832 0.9448383 0.82973004 0.6955746
#> [17,] 0.5395025 0.4890206 0.4012593 0.3833938 0.3584767 0.33653730 0.3119519
#> [18,] 0.3258000 0.2111000 0.2302000 0.3361000 0.4259000 0.46090000 0.2606000
#> [19,] 0.4969508 0.6062380 0.8064632 0.9772698 0.8930166 0.85841660 0.9268146
#> [20,] 0.1965817 0.2444147 0.1243944 0.1296113 0.1647100 0.07790958 0.2365261
#> [21,] 0.2417046 0.3063402 0.4543977 0.6622179 0.6908940 0.78360545 0.8220026
#> [22,] 0.2886820 0.4220296 0.6938060 0.8339046 0.9868082 0.84802601 0.6979728
#> [23,] 0.3187760 0.4398456 0.5548138 0.5762873 0.6161120 0.81372832 0.9828634
#> [24,] 0.3223499 0.2808880 0.3452003 0.4620201 0.5970796 0.63739629 0.6459358
#> [25,] 0.1623246 0.2914093 0.5200421 0.7826780 0.8485659 0.80374457 0.8326003
#>             V27       V28       V29           V3       V30       V31       V32
#>  [1,] 0.9396125 0.7750199 0.6059285  0.058502966 0.4186794 0.3077003 0.2699253
#>  [2,] 0.9766136 1.0047078 0.9134416  0.046606906 0.6934744 0.4652989 0.2786600
#>  [3,] 0.1245726 0.2599600 0.5158855 -0.009014677 0.6204166 0.4199766 0.3774614
#>  [4,] 0.8639365 0.8422712 0.7267978  0.057091412 0.6295687 0.5849341 0.6660872
#>  [5,] 0.9144274 0.9332281 0.8963649  0.020974746 0.9232037 0.9063807 0.8211848
#>  [6,] 0.4657706 0.5140533 0.7885371  0.187369532 0.5445586 0.5484358 0.4429682
#>  [7,] 0.7180082 0.8577569 0.8601957  0.065954071 0.9095913 0.8513045 0.6563945
#>  [8,] 0.6726818 0.9123185 0.8352751  0.021981546 0.5552765 0.2142655 0.3119912
#>  [9,] 0.8102101 0.6416787 0.5469187  0.029365597 0.5114225 0.3652746 0.2644935
#> [10,] 0.9010212 0.9330339 0.8848487  0.031261060 0.7643944 0.6716047 0.5888798
#> [11,] 0.2349411 0.2750172 0.2063395  0.055781848 0.1818612 0.1928919 0.3018010
#> [12,] 0.3921676 0.2855461 0.2954269  0.074388622 0.3464042 0.2779314 0.1268735
#> [13,] 0.2696487 0.5686639 0.5206850  0.059271689 0.7003162 0.3545127 0.4058031
#> [14,] 0.7516583 0.5327316 0.3307236  0.054215470 0.2340805 0.2882113 0.5258907
#> [15,] 0.8899237 0.8244859 0.8887945  0.036701802 0.9211408 0.8587966 0.7357997
#> [16,] 0.6867057 0.5144816 0.3818044  0.114890890 0.4372990 0.4079728 0.1959756
#> [17,] 0.3302211 0.4522399 0.6419506  0.078594195 0.7561143 0.7985400 0.7125879
#> [18,] 0.0874000 0.2862000 0.5606000  0.039000000 0.8344000 0.8096000 0.7250000
#> [19,] 0.9359230 0.8839137 0.8019281 -0.002492215 0.7138014 0.6626388 0.5786950
#> [20,] 0.4312525 0.6089589 0.6533946  0.029860018 0.6449329 0.6536106 0.5407477
#> [21,] 0.6783968 0.6453022 0.5989204  0.012497376 0.5063686 0.4417151 0.4305865
#> [22,] 0.7463827 0.6297761 0.4968414  0.067806913 0.5695997 0.4885563 0.1940670
#> [23,] 0.9821008 0.9154219 0.7402404  0.014088925 0.7752259 0.8554452 0.7431192
#> [24,] 0.5986244 0.7435384 0.9555128  0.014973031 0.9829294 0.9034077 0.8219883
#> [25,] 0.8296150 0.8860793 0.9695549  0.029735340 0.8321613 0.6949933 0.5791990
#>             V33       V34          V35         V36         V37        V38
#>  [1,] 0.1881249 0.1550803  0.174748737  0.24583899 0.263407824 0.29564918
#>  [2,] 0.2476465 0.2356096  0.168294640 -0.01641091 0.006259539 0.05744059
#>  [3,] 0.4678121 0.3672028  0.250972543  0.18789967 0.119484363 0.17143949
#>  [4,] 0.5885640 0.4188120  0.214830183  0.27765831 0.257634659 0.38947448
#>  [5,] 0.7105317 0.6477573  0.646584975  0.69427556 0.669438769 0.55267958
#>  [6,] 0.4357042 0.1773447 -0.000992834  0.44747047 0.580189225 0.24316207
#>  [7,] 0.3038598 0.3177182  0.368635865  0.22928633 0.286637187 0.32532456
#>  [8,] 0.3105464 0.1404414  0.122896607  0.11111483 0.072214371 0.11110430
#>  [9,] 0.2873297 0.3705577  0.462743794  0.48478210 0.394299222 0.26030927
#> [10,] 0.4970567 0.4219126  0.208403155  0.10058668 0.167457764 0.16574231
#> [11,] 0.2872408 0.2684658  0.374629061  0.12204830 0.197008081 0.29315320
#> [12,] 0.3330726 0.5639243  0.649749841  0.58976087 0.481650991 0.38910546
#> [13,] 0.6583602 0.7899471  0.898835103  0.88323013 0.871605035 0.90808544
#> [14,] 0.5673231 0.4389884  0.403071736  0.50981537 0.547233048 0.30587224
#> [15,] 0.7046468 0.7929595  0.925535864  0.99877651 0.920735939 0.76508920
#> [16,] 0.1770510 0.1016382  0.431616731  0.67327143 0.635484308 0.35898633
#> [17,] 0.6137461 0.6625926  0.687125012  0.72482208 0.779792799 0.88426963
#> [18,] 0.8048000 0.9435000  1.000000000  0.89600000 0.551600000 0.30370000
#> [19,] 0.5368090 0.5611846  0.546765089  0.45744336 0.430220412 0.29952570
#> [20,] 0.6477415 0.8110147  0.850623591  0.86441114 0.782269943 0.68154440
#> [21,] 0.5688853 0.6987164  0.724971860  0.59257397 0.282217768 0.17373446
#> [22,] 0.1953263 0.1186807  0.471044805  0.68310199 0.671687208 0.25699012
#> [23,] 0.6726968 0.6100375  0.622565064  0.59403995 0.568927600 0.52242856
#> [24,] 0.7792589 0.6933570  0.603627973  0.49030713 0.428923968 0.25736821
#> [25,] 0.5712503 0.6356785  0.609524204  0.48117015 0.250789013 0.12181953
#>             V39         V4        V40        V41        V42        V43
#>  [1,] 0.3378984 0.06879057 0.29370110 0.26380815 0.24756961 0.25299311
#>  [2,] 0.0832550 0.04209245 0.12454135 0.06026137 0.08528030 0.20190923
#>  [3,] 0.1575662 0.04623364 0.05751990 0.14758075 0.16159601 0.09956293
#>  [4,] 0.5747307 0.04324169 0.49569741 0.41717256 0.52328609 0.45641674
#>  [5,] 0.4701921 0.04765429 0.49545470 0.48244073 0.43255468 0.44282406
#>  [6,] 0.1916543 0.16798002 0.27188036 0.28575252 0.47743215 0.33984808
#>  [7,] 0.2801291 0.04957020 0.20473810 0.57980407 0.62605989 0.54968677
#>  [8,] 0.1115656 0.06145110 0.13236910 0.24736661 0.28619289 0.18654874
#>  [9,] 0.3550680 0.07961293 0.42240034 0.46971771 0.51758552 0.35865421
#> [10,] 0.1699109 0.06563410 0.31250336 0.37261549 0.38867598 0.29753245
#> [11,] 0.2418111 0.07102926 0.07969191 0.02982443 0.24347688 0.21851922
#> [12,] 0.3988720 0.10935453 0.29644807 0.24777776 0.37175699 0.38360395
#> [13,] 0.7187879 0.09385975 0.48090090 0.30497126 0.26967530 0.21871592
#> [14,] 0.4330881 0.05075295 0.44880138 0.32917896 0.30735226 0.25307744
#> [15,] 0.6103037 0.03002546 0.42583068 0.29974505 0.21917811 0.24807144
#> [16,] 0.1237690 0.09989457 0.29845987 0.35995352 0.33387622 0.22860691
#> [17,] 0.8117392 0.05610035 0.68860172 0.62456199 0.50598759 0.43432931
#> [18,] 0.2338000 0.01730000 0.23820000 0.33180000 0.38210000 0.15750000
#> [19,] 0.1275269 0.02189201 0.07445944 0.10045475 0.01232026 0.01739306
#> [20,] 0.6289064 0.03546858 0.56723113 0.44440830 0.36607887 0.26314691
#> [21,] 0.2493167 0.02524835 0.36034929 0.30454262 0.35669283 0.37769538
#> [22,] 0.1806822 0.04543082 0.46837226 0.48730395 0.40682298 0.24969683
#> [23,] 0.2898799 0.02731123 0.15085102 0.19817399 0.20147548 0.21883957
#> [24,] 0.1544344 0.02845692 0.20670235 0.21719186 0.18647693 0.22939438
#> [25,] 0.1442492 0.04228025 0.20532858 0.17245345 0.18372820 0.27244418
#>               V44         V45          V46        V47        V48         V49
#>  [1,]  0.22989718  0.17580692  0.149503556 0.10129024 0.06469912 0.045307578
#>  [2,]  0.21644030  0.19992157  0.216345136 0.17301967 0.10828347 0.062768537
#>  [3,]  0.18051414  0.20099462  0.040119967 0.03143147 0.05768617 0.036339969
#>  [4,]  0.41892934  0.47862799  0.301195731 0.18972574 0.18183732 0.111171330
#>  [5,]  0.51090772  0.47942786  0.344025846 0.24686420 0.19670540 0.105408826
#>  [6,]  0.06769627  0.11048671  0.218921951 0.35512621 0.32481114 0.161026109
#>  [7,]  0.42155107  0.54713747  0.578667461 0.39733289 0.24112474 0.161187044
#>  [8,]  0.25654664  0.20155584  0.177462491 0.09636483 0.03441999 0.041705973
#>  [9,]  0.20673588  0.11872413  0.174483909 0.18609095 0.12066899 0.061267409
#> [10,]  0.15534696  0.20456148  0.234164532 0.22626943 0.18613523 0.095245097
#> [11,]  0.05259034  0.12624821  0.098571018 0.05126157 0.05778024 0.040639662
#> [12,]  0.24775839  0.14083893  0.070174904 0.08562656 0.04278977 0.047996228
#> [13,]  0.27735129  0.42310613  0.304738823 0.14579083 0.10966220 0.058501589
#> [14,]  0.22500042  0.09519144  0.053496096 0.04512241 0.03614826 0.044316387
#> [15,]  0.18837166  0.21813996  0.191341218 0.11012895 0.03973987 0.017091280
#> [16,]  0.20762308  0.21346143  0.139992875 0.14409840 0.14854925 0.092736027
#> [17,]  0.39908988  0.23230904  0.174179156 0.14307222 0.10517950 0.057678536
#> [18,]  0.22280000  0.15820000  0.143300000 0.16340000 0.11330000 0.056700000
#> [19,] -0.03157649 -0.03502582 -0.002883415 0.01526727 0.02943298 0.011702759
#> [20,]  0.16642172  0.12560956  0.073017378 0.07538859 0.07315625 0.037982984
#> [21,]  0.28989054  0.10967317  0.043787330 0.04551211 0.05897551 0.012783278
#> [22,]  0.15447290  0.19414601  0.213699251 0.18589231 0.16079290 0.100196120
#> [23,]  0.17514284  0.13319094  0.136818293 0.07497769 0.04743873 0.032546617
#> [24,]  0.29235738  0.28590786  0.184859297 0.08627604 0.02930154 0.018372118
#> [25,]  0.17175620  0.05255017  0.059133251 0.05038483 0.04206224 0.008061909
#>               V5         V50         V51         V52           V53          V54
#>  [1,] 0.10947271 0.021324145 0.019289534 0.013868802  0.0117357810 0.0208774046
#>  [2,] 0.07332030 0.016155257 0.019440369 0.014053029  0.0077229689 0.0045493570
#>  [3,] 0.05827771 0.008205128 0.018248064 0.011688596 -0.0010394129 0.0026070339
#>  [4,] 0.09431651 0.034892984 0.027629472 0.017990261  0.0071492859 0.0093850215
#>  [5,] 0.07575211 0.029848842 0.017674688 0.013223638  0.0100670978 0.0066730685
#>  [6,] 0.16056128 0.022437451 0.040562584 0.049458478  0.0250922500 0.0313438497
#>  [7,] 0.08388342 0.065193891 0.056586793 0.028006700  0.0163623661 0.0215651970
#>  [8,] 0.10952396 0.003550423 0.004680337 0.012557040  0.0059152381 0.0009228379
#>  [9,] 0.11832276 0.024111484 0.022373632 0.018042285  0.0210560611 0.0134689321
#> [10,] 0.10684241 0.016195036 0.019496865 0.016340152  0.0079093775 0.0053443522
#> [11,] 0.10565940 0.030992522 0.027135497 0.022632922  0.0231541079 0.0210658990
#> [12,] 0.11301092 0.012484172 0.017016781 0.010454526  0.0078146687 0.0257676086
#> [13,] 0.07504126 0.024207207 0.023063670 0.022004998  0.0135226797 0.0241597756
#> [14,] 0.04525295 0.019462295 0.008474842 0.009988030  0.0066641720 0.0209329168
#> [15,] 0.01624568 0.034153584 0.011697070 0.012023928  0.0053853760 0.0129519387
#> [16,] 0.09412851 0.020058462 0.018902178 0.011937864  0.0068329644 0.0159320364
#> [17,] 0.08786973 0.020980999 0.013113299 0.015180110  0.0168655341 0.0056248272
#> [18,] 0.04760000 0.013300000 0.017000000 0.003500000  0.0052000000 0.0083000000
#> [19,] 0.05166197 0.006145933 0.001272837 0.002643340  0.0086939255 0.0066028366
#> [20,] 0.03127189 0.015779676 0.012370777 0.007628168  0.0146381135 0.0122180816
#> [21,] 0.06361518 0.011669279 0.020123466 0.004563663  0.0080902385 0.0125185854
#> [22,] 0.02579587 0.021260066 0.022094771 0.009835097  0.0053531941 0.0090090568
#> [23,] 0.04117740 0.017988996 0.004529665 0.008400000  0.0008728633 0.0018000000
#> [24,] 0.03162338 0.014968282 0.009092049 0.005247782  0.0074365322 0.0102652793
#> [25,] 0.05116954 0.009225853 0.008504583 0.006711781  0.0073528171 0.0052732326
#>                V55           V56          V57         V58         V59
#>  [1,]  0.013821312  0.0105936950  0.009886324 0.015558382 0.011239718
#>  [2,]  0.004954957  0.0037131047  0.005305961 0.003958773 0.001673392
#>  [3,]  0.002188892  0.0015725460  0.001177840 0.002270041 0.004540088
#>  [4,]  0.015217211  0.0075310011  0.008321844 0.008444110 0.007196451
#>  [5,]  0.005745556  0.0070613110  0.006000660 0.004546470 0.002676556
#>  [6,]  0.015808229  0.0179406503  0.019712797 0.021187682 0.022924671
#>  [7,]  0.011001334  0.0135052375  0.009526446 0.018195029 0.010237224
#>  [8,] -0.005078110 -0.0001229542 -0.001142714 0.010141036 0.002627141
#>  [9,]  0.003138220  0.0060938145  0.009654861 0.008503279 0.010371357
#> [10,]  0.006275115  0.0062585308  0.005886034 0.011028968 0.008642678
#> [11,]  0.019912454  0.0099072248  0.001028009 0.016810628 0.013052566
#> [12,]  0.011512508  0.0108762365  0.021256984 0.030980162 0.019858093
#> [13,]  0.021511144  0.0137866871  0.011586341 0.015656308 0.014848204
#> [14,]  0.019206312  0.0101402472  0.016233928 0.018022590 0.011590168
#> [15,]  0.010180123  0.0037864494  0.009730712 0.006615983 0.006793455
#> [16,]  0.010633684  0.0059034013  0.008098775 0.009341997 0.012569328
#> [17,]  0.006025576  0.0070045745  0.004280417 0.015131746 0.010875293
#> [18,]  0.007800000  0.0075000000  0.010500000 0.016000000 0.009500000
#> [19,]  0.006702448  0.0029750001  0.002997421 0.001631317 0.002683564
#> [20,]  0.011845513  0.0086180527  0.007979041 0.007377083 0.011179783
#> [21,]  0.001988352  0.0110357334  0.011656116 0.004195064 0.007724222
#> [22,]  0.009152397  0.0033670183  0.015759526 0.007291013 0.009534260
#> [23,]  0.006846617  0.0038731600  0.012153977 0.013373754 0.007052267
#> [24,]  0.003753256  0.0035215723  0.005240844 0.003492187 0.006196121
#> [25,]  0.004944962  0.0055356536  0.011999301 0.010442905 0.007456136
#>               V6           V60         V7         V8         V9
#>  [1,] 0.15059963  0.0073737627 0.18194943 0.24028191 0.31637309
#>  [2,] 0.12867243  0.0042591388 0.15851300 0.19157264 0.25154928
#>  [3,] 0.16221334 -0.0003152367 0.17276463 0.13545940 0.18338336
#>  [4,] 0.11711317  0.0043517972 0.11310188 0.15570855 0.23606775
#>  [5,] 0.10972818  0.0030709094 0.13847751 0.13921525 0.27604475
#>  [6,] 0.31187966  0.0058191788 0.28060700 0.18123750 0.04236253
#>  [7,] 0.11629433  0.0099920290 0.15324710 0.19581019 0.25607737
#>  [8,] 0.14344557  0.0002376930 0.15059968 0.19234542 0.24584373
#>  [9,] 0.06260282  0.0071047862 0.10377098 0.11123954 0.07153278
#> [10,] 0.10255530  0.0044920607 0.07727528 0.11865375 0.16923544
#> [11,] 0.11796758  0.0110193931 0.12157581 0.05582171 0.06792836
#> [12,] 0.12162038  0.0084775247 0.13481418 0.09078174 0.12054285
#> [13,] 0.12579810  0.0137045624 0.25757633 0.44752844 0.59788612
#> [14,] 0.06126120  0.0055305134 0.05918041 0.04608735 0.08011906
#> [15,] 0.06168691  0.0087859270 0.13400052 0.12166734 0.07008782
#> [16,] 0.12253576  0.0082822139 0.19317776 0.20686955 0.24783691
#> [17,] 0.14153371  0.0075403620 0.14489924 0.17204455 0.26486355
#> [18,] 0.08160000  0.0011000000 0.09930000 0.03150000 0.07360000
#> [19,] 0.09593881  0.0030323961 0.10165509 0.08311935 0.14086265
#> [20,] 0.07412444  0.0045528623 0.05570661 0.06696941 0.03401407
#> [21,] 0.08481103  0.0056137125 0.10826667 0.10088908 0.11483702
#> [22,] 0.07862295  0.0081080663 0.09560548 0.22111090 0.28521082
#> [23,] 0.07627635  0.0088875831 0.10350691 0.11477754 0.07865978
#> [24,] 0.02715085  0.0022685811 0.06973693 0.06024897 0.06462122
#> [25,] 0.08302067  0.0057545299 0.06185250 0.10187667 0.09609311
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2463768