Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

r format_bib("kohonen_1990", "kohonen_1995")

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11         V12        V13        V14
#>  [1,] 0.018000000 0.29180000 0.32730000  0.30350000 0.30330000 0.25870000
#>  [2,] 0.029679460 0.33965827 0.39298284  0.35667585 0.23679860 0.05771178
#>  [3,] 0.055986307 0.33069651 0.24276660  0.18319632 0.40118689 0.52284616
#>  [4,] 0.025003125 0.39747678 0.45654468  0.41735126 0.34788777 0.25491496
#>  [5,] 0.018165530 0.13655657 0.24685281  0.29791708 0.23534381 0.20600136
#>  [6,] 0.032421142 0.15594173 0.15913833  0.15610744 0.20470138 0.29580491
#>  [7,] 0.035578704 0.31247201 0.39086983  0.35747651 0.41021159 0.57851246
#>  [8,] 0.022084000 0.21565155 0.10072503 -0.01853569 0.15268212 0.29888968
#>  [9,] 0.048198135 0.25579743 0.35129196  0.42763339 0.52706460 0.48986786
#> [10,] 0.026602757 0.07681646 0.17200695  0.20965788 0.18383795 0.13774676
#> [11,] 0.094963828 0.61441427 0.47775991  0.21849873 0.14760329 0.33765411
#> [12,] 0.047520152 0.18006211 0.21331694  0.23084629 0.28147846 0.28929782
#> [13,] 0.025569955 0.16884737 0.23491149  0.28594271 0.29688301 0.33795953
#> [14,] 0.023789818 0.11143033 0.07256381  0.05173972 0.15360544 0.32058059
#> [15,] 0.032059752 0.04161434 0.07771186  0.23786391 0.13251251 0.11388617
#> [16,] 0.009768789 0.17763672 0.13601552  0.19067159 0.26709158 0.14895872
#> [17,] 0.014316057 0.10780863 0.13183832  0.11866514 0.10955932 0.13971855
#> [18,] 0.012600000 0.42840000 0.30150000  0.12070000 0.32990000 0.57070000
#> [19,] 0.010934198 0.12974363 0.20364311  0.34901745 0.34878417 0.25414195
#> [20,] 0.013530884 0.07384394 0.06999121  0.07158963 0.09165451 0.10427220
#> [21,] 0.008957102 0.31178010 0.35473075  0.37362930 0.40122222 0.45587251
#> [22,] 0.013285288 0.07615833 0.04542147  0.11434611 0.15000347 0.12831889
#> [23,] 0.016644907 0.25790519 0.23558814  0.05926198 0.12437842 0.42652389
#> [24,] 0.042485566 0.22080193 0.25103323  0.32497989 0.27567420 0.28668569
#> [25,] 0.011079126 0.08601272 0.06840957  0.08309783 0.21902561 0.23925671
#>              V15        V16       V17        V18        V19         V2
#>  [1,] 0.16820000 0.13080000 0.2803000 0.45190000 0.66410000 0.04440000
#>  [2,] 0.11078461 0.16311479 0.1948568 0.26075385 0.24343733 0.03301466
#>  [3,] 0.21462652 0.37190141 0.5473154 0.57085391 0.78489173 0.05625321
#>  [4,] 0.18253070 0.22556744 0.2722697 0.36597261 0.50540101 0.03795465
#>  [5,] 0.14359705 0.14443638 0.1883383 0.24566637 0.38849435 0.05540687
#>  [6,] 0.43646065 0.59042553 0.7016633 0.73525814 0.81524297 0.03434715
#>  [7,] 0.62046769 0.57152867 0.5795231 0.71291150 0.84903631 0.03274140
#>  [8,] 0.49254910 0.66168011 0.7516836 0.82728003 0.96156780 0.03620446
#>  [9,] 0.60260357 0.69649500 0.8817485 0.94529033 0.90226414 0.09290688
#> [10,] 0.09374873 0.13698566 0.1667172 0.28736429 0.33689549 0.05455879
#> [11,] 0.47602059 0.47223210 0.2757619 0.27427174 0.46249575 0.09353573
#> [12,] 0.30567625 0.35670418 0.3758535 0.24765560 0.29783625 0.07905434
#> [13,] 0.40676962 0.51408850 0.5931117 0.68727902 0.78388242 0.02318292
#> [14,] 0.53690939 0.75199932 0.8077557 0.79831934 0.84287545 0.01918192
#> [15,] 0.10944465 0.16117020 0.1389272 0.09786492 0.16801817 0.03527150
#> [16,] 0.20897577 0.23991285 0.2265894 0.26994184 0.25951415 0.04315546
#> [17,] 0.10529451 0.11733897 0.2128092 0.29968609 0.33013496 0.01335509
#> [18,] 0.69620000 0.97510000 1.0000000 0.92930000 0.62100000 0.05190000
#> [19,] 0.23502874 0.54714788 0.7416250 0.75647471 0.60111016 0.02484536
#> [20,] 0.08146388 0.14011965 0.1673422 0.26693946 0.31743067 0.01377350
#> [21,] 0.61139296 0.76592844 0.9301606 0.93683716 0.89164132 0.01490257
#> [22,] 0.14054955 0.09410478 0.0946223 0.16486484 0.09949253 0.03267233
#> [23,] 0.58098876 0.72798988 0.8397796 0.78668247 0.54248677 0.02177698
#> [24,] 0.38717684 0.48259515 0.5004911 0.60553876 0.75769000 0.06668147
#> [25,] 0.21988096 0.26975545 0.2196542 0.31552303 0.60318401 0.02477373
#>             V20       V21        V22        V23       V24        V25       V26
#>  [1,] 0.7683000 0.6960000 0.43930000 0.24320000 0.2886000 0.49740000 0.8172000
#>  [2,] 0.2698129 0.4138441 0.66252341 0.77271751 0.8124579 0.92724756 1.0186445
#>  [3,] 0.8536101 0.5755304 0.67355518 0.84367354 0.6916991 0.76378975 0.8164856
#>  [4,] 0.6240669 0.7808621 0.87850510 0.84773694 0.8755266 0.86039241 0.8400454
#>  [5,] 0.3128081 0.3561118 0.41660336 0.50475302 0.6219994 0.61772170 0.6581505
#>  [6,] 0.9014795 0.9348822 0.95633558 0.83041050 0.6643314 0.52258827 0.4462791
#>  [7,] 0.7589622 0.7606547 0.92492070 0.95472390 0.9058529 0.75885722 0.8363515
#>  [8,] 0.9934255 1.0093065 1.00735719 0.92697081 0.7012863 0.52518097 0.4985832
#>  [9,] 0.7991459 0.7568835 0.52869201 0.28558568 0.2722001 0.24805172 0.1876493
#> [10,] 0.5418773 0.5952389 0.68664552 0.73363990 0.5888398 0.61821073 0.7850761
#> [11,] 0.8842018 0.9510137 0.40536356 0.10893327 0.1029806 0.21492783 0.3809991
#> [12,] 0.3528185 0.4262916 0.48402260 0.50052765 0.5542362 0.56513219 0.5965578
#> [13,] 0.8841387 0.9230928 0.95449757 0.93316231 0.9320691 0.91469095 0.8588883
#> [14,] 0.8696796 0.9710813 0.94940104 0.93180507 0.9695055 0.93910980 0.9353780
#> [15,] 0.1335153 0.2026109 0.02925113 0.22738042 0.2140387 0.01197183 0.1819475
#> [16,] 0.1215944 0.1655151 0.15959157 0.23220767 0.2290099 0.25784970 0.3032770
#> [17,] 0.4423998 0.5521608 0.72768873 0.83676019 0.8377598 0.84584529 0.8811193
#> [18,] 0.4586000 0.5001000 0.50320000 0.70820000 0.8420000 0.81090000 0.7690000
#> [19,] 0.4343045 0.4465195 0.46212186 0.57327469 0.7368706 0.86576300 0.9161017
#> [20,] 0.3274276 0.3272812 0.35104547 0.35768153 0.4484162 0.40954923 0.4815784
#> [21,] 0.7948810 0.7439526 0.80702149 0.57556754 0.3820360 0.53543910 0.6567884
#> [22,] 0.1698416 0.2386652 0.18855953 0.09191132 0.1591535 0.19223122 0.3216413
#> [23,] 0.3553468 0.5241736 0.77813764 0.85150826 0.8393180 0.72970954 0.6955743
#> [24,] 0.8716631 0.9355184 0.84330369 0.70861462 0.6781444 0.81540170 0.8885982
#> [25,] 0.7687318 0.7148724 0.63773497 0.51458831 0.5241468 0.65107685 0.7605964
#>             V27       V28       V29         V3        V30       V31       V32
#>  [1,] 1.0000000 0.9238000 0.8519000 0.04760000 0.77220000 0.5772000 0.5190000
#>  [2,] 1.0049528 0.9195130 0.7902704 0.03596837 0.70646785 0.7368617 0.5969236
#>  [3,] 0.7266219 0.6462702 0.3954269 0.06344272 0.65967249 0.7316496 0.8243796
#>  [4,] 0.8223241 0.7568481 0.5940072 0.03942534 0.46480185 0.3928063 0.3766774
#>  [5,] 0.8387369 0.9425503 0.9103223 0.07391955 0.71624207 0.6330389 0.5392429
#>  [6,] 0.3563623 0.3128289 0.3752244 0.02487801 0.44827203 0.3512237 0.3679804
#>  [7,] 0.8483726 0.6258175 0.5422461 0.05843785 0.48555619 0.4123966 0.2527432
#>  [8,] 0.4888538 0.6062577 0.7003480 0.04585305 0.60168302 0.3360922 0.3123497
#>  [9,] 0.1806601 0.1922140 0.1261287 0.10697955 0.16880206 0.2684300 0.3388482
#> [10,] 0.9586543 0.9731488 0.7376923 0.03617769 0.57195748 0.3804009 0.1862281
#> [11,] 0.2322472 0.5001088 0.6062582 0.05197738 0.77603746 0.4141068 0.3983369
#> [12,] 0.6929561 0.8105053 0.8396215 0.08145459 0.86124976 0.8721824 0.7898348
#> [13,] 0.8210633 0.7654470 0.6416859 0.03448935 0.50332827 0.3635865 0.2383263
#> [14,] 0.9058360 0.6971439 0.4021586 0.03316053 0.34009714 0.4164462 0.3927690
#> [15,] 0.4649117 0.6885789 0.6736628 0.03580496 0.63693273 0.7083608 0.5347003
#> [16,] 0.1737268 0.4873910 0.6280215 0.03529830 0.58371874 0.5014245 0.6245497
#> [17,] 0.8242227 0.7806802 0.7614677 0.02816218 0.68786308 0.5957203 0.5666072
#> [18,] 0.8105000 0.6203000 0.2356000 0.06210000 0.25950000 0.6299000 0.6762000
#> [19,] 0.9517666 0.9409388 0.9142971 0.02100015 0.84026845 0.7330118 0.6729052
#> [20,] 0.5922323 0.7893270 0.9416989 0.02119579 0.86968168 0.6947724 0.5539378
#> [21,] 0.6187339 0.5112546 0.5337870 0.03280765 0.62111374 0.5571929 0.3558693
#> [22,] 0.3906679 0.4860910 0.5510324 0.01714522 0.63343101 0.7029193 0.7137947
#> [23,] 0.6471727 0.4936514 0.3574306 0.06253758 0.44817790 0.3917810 0.2818198
#> [24,] 0.7559075 0.5714353 0.3083718 0.03521092 0.05051646 0.2482066 0.7382339
#> [25,] 0.8562018 0.7459846 0.5104955 0.02359119 0.38620220 0.3837076 0.3317634
#>             V33        V34        V35        V36        V37        V38
#>  [1,] 0.6824000 0.62200000 0.50540000 0.35780000 0.38090000 0.38130000
#>  [2,] 0.3832696 0.17060186 0.09796198 0.04793453 0.14015039 0.15029900
#>  [3,] 0.7297936 0.60749660 0.36875960 0.13177114 0.10234912 0.38811818
#>  [4,] 0.4185304 0.40473306 0.28964330 0.24725930 0.23845885 0.25319452
#>  [5,] 0.4300237 0.31044115 0.09769902 0.02867325 0.20438225 0.18813453
#>  [6,] 0.4547559 0.37919742 0.28837353 0.30159235 0.30706533 0.32086062
#>  [7,] 0.1116598 0.08235528 0.15997541 0.16591618 0.11475498 0.13898203
#>  [8,] 0.4557910 0.37822844 0.22181768 0.07355879 0.02588885 0.06574002
#>  [9,] 0.3699484 0.45052770 0.53042094 0.20811968 0.12819730 0.30291615
#> [10,] 0.1630672 0.41561062 0.66614302 0.76602586 0.77068539 0.67733999
#> [11,] 0.5906779 0.76142275 0.84917836 0.79420009 0.77742536 0.86420672
#> [12,] 0.5848554 0.52998330 0.58236687 0.42051986 0.35095480 0.39566479
#> [13,] 0.2241659 0.18439962 0.21717710 0.20476842 0.24624682 0.28334436
#> [14,] 0.3431367 0.27739623 0.20663256 0.09427751 0.20856760 0.18509999
#> [15,] 0.6465417 0.87599773 0.96291613 0.97087407 0.77376290 0.66626642
#> [16,] 0.7029159 0.57705351 0.54213185 0.67558274 0.81570489 0.91891556
#> [17,] 0.6124610 0.67013111 0.67781662 0.60425794 0.47186414 0.34996514
#> [18,] 0.2903000 0.43930000 0.85290000 0.71800000 0.48010000 0.58560000
#> [19,] 0.7485661 0.85530716 0.95851877 0.95112866 0.90030310 0.80269251
#> [20,] 0.6417194 0.57580644 0.43619859 0.44119861 0.45949475 0.42982434
#> [21,] 0.3816582 0.36634155 0.21254147 0.17530258 0.23793125 0.23468509
#> [22,] 0.7537190 0.87338801 0.99511713 0.97977709 0.85358877 0.68625886
#> [23,] 0.2018995 0.06238654 0.37511211 0.62437204 0.58887370 0.25685006
#> [24,] 0.7926275 0.82212710 0.80832902 0.80073684 0.65954773 0.13344371
#> [25,] 0.3455563 0.37218110 0.36704212 0.39014812 0.45807399 0.45360375
#>              V39         V4        V40        V41        V42        V43
#>  [1,] 0.33590000 0.06980000 0.27710000 0.36480000 0.38340000 0.34530000
#>  [2,] 0.14994462 0.06545943 0.07898343 0.10569913 0.09984962 0.15029799
#>  [3,] 0.60908858 0.09889732 0.44019759 0.46410511 0.49247605 0.42208509
#>  [4,] 0.35008478 0.03971631 0.36385905 0.20898484 0.23837301 0.31643895
#>  [5,] 0.19718778 0.11007987 0.27916591 0.30092261 0.34837789 0.22426407
#>  [6,] 0.30652569 0.04140920 0.25888189 0.25374345 0.26679101 0.18386765
#>  [7,] 0.23361684 0.08973480 0.25924578 0.27973987 0.25493809 0.17397132
#>  [8,] 0.01741414 0.05773202 0.09215491 0.21417170 0.22569442 0.09301442
#>  [9,] 0.31483389 0.13544790 0.15495633 0.06815106 0.08497783 0.09634227
#> [10,] 0.60740980 0.04357319 0.44773363 0.36308395 0.30674587 0.37197071
#> [11,] 0.78134109 0.12124793 0.56843135 0.37019080 0.35085296 0.33975211
#> [12,] 0.32894152 0.08661988 0.24181023 0.55191011 0.61495266 0.46235510
#> [13,] 0.27862283 0.09149222 0.26620904 0.29667417 0.32801822 0.27498729
#> [14,] 0.12704156 0.03388639 0.14721822 0.09407585 0.05483913 0.07567636
#> [15,] 0.62942496 0.02643898 0.66236029 0.59089522 0.51109353 0.37070770
#> [16,] 0.97505366 0.02776780 0.92887908 0.85157269 0.74288265 0.52006676
#> [17,] 0.23146378 0.03336133 0.26530349 0.25832398 0.21535939 0.21894546
#> [18,] 0.49930000 0.05180000 0.28660000 0.06010000 0.11670000 0.27370000
#> [19,] 0.61456496 0.03231536 0.43538758 0.27309257 0.17307198 0.11535167
#> [20,] 0.39457698 0.02920250 0.30260320 0.17845820 0.11077544 0.11742785
#> [21,] 0.27428520 0.05385056 0.32103078 0.28844257 0.22451970 0.11520733
#> [22,] 0.59718111 0.03022506 0.50812243 0.39998309 0.35218715 0.25861275
#> [23,] 0.15386918 0.04122987 0.40248742 0.44419020 0.30631373 0.18410825
#> [24,] 0.33221282 0.07198607 0.56317089 0.55029162 0.49216509 0.39359026
#> [25,] 0.43726016 0.02174191 0.34072669 0.24513260 0.17974348 0.13407953
#>              V44        V45        V46        V47        V48        V49
#>  [1,] 0.20960000 0.10310000 0.07980000 0.07010000 0.05260000 0.02410000
#>  [2,] 0.15599075 0.11691143 0.08677764 0.11299913 0.10758928 0.06956316
#>  [3,] 0.36959995 0.49349138 0.50958736 0.29630129 0.22957261 0.12836620
#>  [4,] 0.31369362 0.28323148 0.18439798 0.14633899 0.12745270 0.07132441
#>  [5,] 0.13939552 0.19134290 0.21734378 0.25306702 0.21428425 0.09946151
#>  [6,] 0.19081276 0.16571214 0.07897314 0.06505625 0.05280823 0.03602962
#>  [7,] 0.14747553 0.12867014 0.12782098 0.07385770 0.03584122 0.02888641
#>  [8,] 0.15803591 0.17653704 0.12580998 0.11644353 0.08327446 0.06079355
#>  [9,] 0.05654843 0.05095147 0.05365918 0.08676389 0.08799193 0.05702124
#> [10,] 0.33564296 0.42049964 0.36568327 0.23182120 0.11837263 0.08834106
#> [11,] 0.34936830 0.48761935 0.38652320 0.12314469 0.13179752 0.06977742
#> [12,] 0.45802536 0.55881997 0.51671726 0.31880686 0.17708503 0.09992357
#> [13,] 0.17731047 0.14842330 0.09052519 0.08573717 0.08106248 0.05271164
#> [14,] 0.13138557 0.11305040 0.05410618 0.06088018 0.05059881 0.02740897
#> [15,] 0.21312691 0.13673761 0.06958261 0.08015068 0.07697571 0.05543529
#> [16,] 0.29000208 0.43971251 0.52035317 0.40556350 0.26835588 0.17645654
#> [17,] 0.15223987 0.09083811 0.06267638 0.04379615 0.04220561 0.01832512
#> [18,] 0.28120000 0.20780000 0.06600000 0.04910000 0.03450000 0.01720000
#> [19,] 0.06474061 0.14227067 0.18442754 0.10590963 0.06211343 0.01609378
#> [20,] 0.12907107 0.12431876 0.07204913 0.05432391 0.05195393 0.02288671
#> [21,] 0.08329957 0.15774176 0.17529541 0.16659206 0.09500699 0.03776185
#> [22,] 0.21183430 0.18291215 0.12123301 0.10002611 0.08650473 0.03950134
#> [23,] 0.16020007 0.15368799 0.09459808 0.13526082 0.13984867 0.08587486
#> [24,] 0.39325236 0.15468867 0.03924808 0.06765024 0.04127382 0.04779340
#> [25,] 0.09145681 0.09560216 0.10263745 0.07546645 0.04877836 0.02729566
#>               V5         V50         V51         V52         V53         V54
#>  [1,] 0.16150000 0.011700000 0.012200000 0.012200000 0.011400000 0.009800000
#>  [2,] 0.05654707 0.024877062 0.011409005 0.010294582 0.004815464 0.010717775
#>  [3,] 0.21503159 0.076134733 0.020158742 0.018450059 0.029677216 0.019912122
#>  [4,] 0.05577207 0.019724733 0.018363799 0.011354259 0.009310335 0.011516641
#>  [5,] 0.10970648 0.015725915 0.024741133 0.021555875 0.014444728 0.012462418
#>  [6,] 0.07163027 0.013750552 0.013667061 0.010082100 0.006117603 0.009198733
#>  [7,] 0.11407934 0.016088285 0.021400266 0.022023375 0.008057053 0.025592811
#>  [8,] 0.10668509 0.018214085 0.006073416 0.010451365 0.006449566 0.001062892
#>  [9,] 0.11143329 0.023126501 0.024142322 0.028644540 0.015251567 0.013201950
#> [10,] 0.02763035 0.029799480 0.012972789 0.014917618 0.006736357 0.001671749
#> [11,] 0.10135969 0.039071280 0.018907135 0.026231893 0.025647819 0.021321119
#> [12,] 0.08704456 0.039310678 0.041771270 0.032836922 0.019150946 0.019099306
#> [13,] 0.15095592 0.021691520 0.018166595 0.019974515 0.018217039 0.014567867
#> [14,] 0.07080806 0.009875533 0.014802184 0.012219725 0.010067010 0.005562464
#> [15,] 0.06629590 0.017559239 0.005771990 0.007703649 0.016452036 0.016398978
#> [16,] 0.05717603 0.069857924 0.037012877 0.016450611 0.012986183 0.012077725
#> [17,] 0.05811375 0.010847527 0.008521588 0.008006838 0.010424765 0.008202003
#> [18,] 0.10720000 0.028700000 0.002700000 0.020800000 0.004800000 0.019900000
#> [19,] 0.02988227 0.020075230 0.027934526 0.019513272 0.007782591 0.004989268
#> [20,] 0.02980356 0.009768656 0.006099182 0.008057499 0.012853227 0.009933927
#> [21,] 0.13637384 0.017707723 0.006658381 0.005448364 0.003473365 0.004374426
#> [22,] 0.01456579 0.016344034 0.019598550 0.007149283 0.009411673 0.005812959
#> [23,] 0.07464357 0.013273271 0.024436101 0.010500856 0.010084393 0.008996273
#> [24,] 0.07217853 0.022258225 0.006727053 0.007805520 0.013465878 0.018778956
#> [25,] 0.02455537 0.021504526 0.015093033 0.012735345 0.014339364 0.018857671
#>                V55         V56         V57         V58         V59         V6
#>  [1,] 0.0027000000 0.002500000 0.002600000 0.005000000 0.007300000 0.08870000
#>  [2,] 0.0032289312 0.005990449 0.004500258 0.007396560 0.008710289 0.13323157
#>  [3,] 0.0377212134 0.032229489 0.030304255 0.037634733 0.021232401 0.12928526
#>  [4,] 0.0076031393 0.005316110 0.003368321 0.006154546 0.006097821 0.08238309
#>  [5,] 0.0119265296 0.009604911 0.010309953 0.010941581 0.008257329 0.11243479
#>  [6,] 0.0046004221 0.007902946 0.006145912 0.005815626 0.007584111 0.12350391
#>  [7,] 0.0124201306 0.010545375 0.011386642 0.011735302 0.011155810 0.16442291
#>  [8,] 0.0004371252 0.005824260 0.003785679 0.010773846 0.006829013 0.14973460
#>  [9,] 0.0180362284 0.011864623 0.003591900 0.016326259 0.019573581 0.02908567
#> [10,] 0.0085521343 0.004384352 0.003920095 0.004729282 0.005782932 0.05919519
#> [11,] 0.0181972084 0.017561679 0.013626539 0.011288311 0.006419222 0.07869986
#> [12,] 0.0148706878 0.011338631 0.011721082 0.012498834 0.014209018 0.10656841
#> [13,] 0.0042924423 0.010286598 0.007666086 0.007413131 0.008162668 0.14189517
#> [14,] 0.0038743665 0.008789316 0.006606293 0.003955222 0.006154459 0.11595330
#> [15,] 0.0081329719 0.006479011 0.012936052 0.009287191 0.012563430 0.10909002
#> [16,] 0.0182631631 0.012785341 0.003194684 0.014035173 0.009392328 0.09533042
#> [17,] 0.0058865753 0.008531399 0.008911109 0.004927070 0.004904237 0.08413142
#> [18,] 0.0126000000 0.002200000 0.003700000 0.003400000 0.011400000 0.25870000
#> [19,] 0.0084951106 0.008830378 0.007701756 0.012095092 0.006277786 0.11209379
#> [20,] 0.0081595505 0.006247216 0.005567483 0.003478039 0.008029884 0.05651661
#> [21,] 0.0091392938 0.008410431 0.010822258 0.005124967 0.004533732 0.15626063
#> [22,] 0.0146701811 0.010031039 0.004030996 0.005489021 0.005259155 0.06801913
#> [23,] 0.0050501993 0.007202380 0.010267068 0.013737908 0.014780493 0.09298853
#> [24,] 0.0075142041 0.018112579 0.018453550 0.011038859 0.006255532 0.19668749
#> [25,] 0.0129196894 0.009029259 0.007127302 0.010475554 0.004220127 0.02432346
#>                V60         V7         V8         V9
#>  [1,]  0.002200000 0.05960000 0.10710000 0.31750000
#>  [2,]  0.010195112 0.14840516 0.16431925 0.24024525
#>  [3,]  0.008553788 0.17375653 0.10839038 0.24565490
#>  [4,]  0.004774367 0.11681547 0.15949925 0.28072799
#>  [5,]  0.005472293 0.09843494 0.07611126 0.10694020
#>  [6,]  0.006330315 0.13762087 0.10837734 0.16110498
#>  [7,]  0.007172307 0.24631708 0.30576577 0.32225373
#>  [8,] -0.001299251 0.15810328 0.16083870 0.19423473
#>  [9,]  0.012666713 0.08159635 0.04131207 0.11833752
#> [10,]  0.005681266 0.03521133 0.06116611 0.03389982
#> [11,]  0.011687567 0.16662326 0.32016681 0.55470872
#> [12,]  0.010539475 0.09809535 0.13502512 0.22126731
#> [13,]  0.004973666 0.12871463 0.13471806 0.13502366
#> [14,]  0.004480756 0.12773639 0.09201368 0.12736040
#> [15,]  0.005893203 0.09397972 0.06111213 0.01042267
#> [16,]  0.012995623 0.21484596 0.22025466 0.01595000
#> [17,]  0.007547200 0.09426955 0.09300402 0.09595409
#> [18,]  0.007700000 0.23040000 0.20670000 0.34160000
#> [19,]  0.011119446 0.14041550 0.06535180 0.06414266
#> [20,]  0.005492565 0.04309809 0.05098748 0.07250107
#> [21,]  0.005294177 0.14416952 0.17635059 0.26637295
#> [22,]  0.001127110 0.06777634 0.06643107 0.01236793
#> [23,]  0.012412686 0.13224202 0.19324953 0.19596773
#> [24,]  0.004602559 0.10274325 0.09905709 0.04100645
#> [25,]  0.003717558 0.03758797 0.03974216 0.07777586
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913