Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11         V12        V13        V14
#>  [1,] 0.005175390 0.17650703 0.19055344 0.134292650 0.03673986 0.13267256
#>  [2,] 0.021445263 0.19569223 0.32586499 0.327686150 0.29178375 0.24820597
#>  [3,] 0.025982812 0.38089802 0.38217156 0.325515732 0.28327777 0.16503099
#>  [4,] 0.035219883 0.28482231 0.31807138 0.321371780 0.22363878 0.15487502
#>  [5,] 0.015980497 0.24456787 0.21508763 0.254519572 0.26025225 0.22164687
#>  [6,] 0.043186594 0.18572426 0.28107671 0.290753449 0.25744934 0.25467708
#>  [7,] 0.030063682 0.36698870 0.42808464 0.461342984 0.46090152 0.41175377
#>  [8,] 0.024653793 0.21208437 0.20056623 0.163758433 0.14193253 0.11701943
#>  [9,] 0.052845292 0.17362660 0.26934626 0.314745220 0.34192729 0.32007983
#> [10,] 0.059651583 0.09420773 0.13376355 0.161422698 0.10072318 0.19704976
#> [11,] 0.059570155 0.24664128 0.34415609 0.445802376 0.51676137 0.58668912
#> [12,] 0.024418921 0.40139942 0.33448725 0.260708021 0.24760263 0.24459765
#> [13,] 0.055854024 0.14154001 0.21509738 0.304473302 0.31222288 0.31632440
#> [14,] 0.009910751 0.06423586 0.06699558 0.077859639 0.10572845 0.14244375
#> [15,] 0.022367653 0.07390537 0.09971390 0.181965115 0.14373336 0.17768752
#> [16,] 0.027461684 0.30149211 0.30206034 0.249625145 0.30891691 0.50320070
#> [17,] 0.008668615 0.13083959 0.12060676 0.007436369 0.04854307 0.06302191
#> [18,] 0.012706832 0.06923776 0.06761448 0.118040138 0.11506830 0.09762549
#> [19,] 0.009148371 0.19113271 0.12720901 0.171352015 0.35012228 0.24268116
#> [20,] 0.015226047 0.16181723 0.18924724 0.120689350 0.09485600 0.14599770
#> [21,] 0.014518055 0.12165067 0.14088319 0.145035677 0.17077975 0.16930323
#> [22,] 0.016725377 0.14715209 0.15403518 0.123461030 0.16446092 0.16757051
#> [23,] 0.013338194 0.17791425 0.12814216 0.078627232 0.09500735 0.10565370
#> [24,] 0.013314911 0.11631463 0.07406805 0.060826832 0.07060342 0.14653664
#> [25,] 0.014420231 0.13661685 0.12120475 0.084834559 0.06301324 0.12107594
#>              V15        V16       V17       V18       V19          V2       V20
#>  [1,] 0.25016630 0.54869639 0.7043923 0.8069901 0.8438884 0.009485004 0.7140278
#>  [2,] 0.24687088 0.10967983 0.1836879 0.2422133 0.2533924 0.016526789 0.3030037
#>  [3,] 0.17478783 0.20654822 0.2180220 0.2704349 0.2376721 0.037341764 0.1824823
#>  [4,] 0.14823136 0.18451804 0.1951514 0.3187140 0.4778150 0.051016460 0.6344362
#>  [5,] 0.15698638 0.13041262 0.1328104 0.2359374 0.4304799 0.038959997 0.4607490
#>  [6,] 0.30778766 0.40035151 0.5262887 0.6912171 0.8023296 0.041552376 0.8530680
#>  [7,] 0.35231921 0.34614617 0.4293357 0.5011373 0.6252298 0.035167927 0.7445532
#>  [8,] 0.27414659 0.44306162 0.5367969 0.5227970 0.6023395 0.033425265 0.7809671
#>  [9,] 0.26377461 0.26286498 0.3183104 0.2447509 0.2594890 0.070600495 0.4504521
#> [10,] 0.29095504 0.47750338 0.6404362 0.6679586 0.8129011 0.058374689 0.9150736
#> [11,] 0.69814962 0.76678843 0.8404728 0.8837320 0.9342945 0.078909004 0.9065255
#> [12,] 0.20471931 0.27953175 0.2641176 0.2926898 0.2642047 0.032101405 0.3730021
#> [13,] 0.19288076 0.05870555 0.1363136 0.1971248 0.2619222 0.067813914 0.2972758
#> [14,] 0.16919446 0.24918637 0.3102341 0.3446908 0.3612394 0.013564820 0.3731796
#> [15,] 0.20458798 0.37189216 0.6933493 0.5816944 0.2580599 0.027794121 0.1960938
#> [16,] 0.63991354 0.72905398 0.8104081 0.7656779 0.6017496 0.029050405 0.5305953
#> [17,] 0.12478605 0.06129495 0.1042378 0.2132931 0.3467822 0.017047494 0.3698795
#> [18,] 0.07466006 0.12827249 0.1288942 0.1861724 0.1909083 0.019732965 0.1954458
#> [19,] 0.05527997 0.20671169 0.2921789 0.3573457 0.6452539 0.042360510 0.7365263
#> [20,] 0.31172150 0.25351563 0.1610662 0.1610720 0.1866752 0.018059112 0.2842181
#> [21,] 0.21981526 0.28926742 0.3239041 0.3209276 0.2404683 0.017878331 0.2172953
#> [22,] 0.14486832 0.12749516 0.1103654 0.1757736 0.3550759 0.016166268 0.3425762
#> [23,] 0.14906308 0.23986354 0.2667125 0.2472982 0.3532646 0.002990906 0.5590644
#> [24,] 0.07261289 0.06193040 0.1953223 0.3054840 0.2983856 0.013651238 0.4551708
#> [25,] 0.09963392 0.10865816 0.2276178 0.2676817 0.3283185 0.002796684 0.5091242
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.7460839 0.8992924 0.9112293 0.4802467 0.1377725 0.2926821 0.6322598
#>  [2,] 0.4179320 0.4215128 0.5089544 0.5935906 0.7241519 0.8086268 0.9687322
#>  [3,] 0.2992168 0.6273241 0.8030036 0.8333048 0.8508592 0.8609418 0.9285280
#>  [4,] 0.7788756 0.8460197 0.8142290 0.8654521 0.9072039 0.9636647 0.9497755
#>  [5,] 0.5052185 0.6428497 0.6243084 0.6352427 0.7293542 0.8232146 0.9451478
#>  [6,] 0.8908403 0.9502379 0.9542989 0.9242938 0.8102944 0.6329203 0.4643391
#>  [7,] 0.8161286 0.8915184 0.8769232 0.8896448 0.8878919 0.8811731 0.8923132
#>  [8,] 0.9276462 1.0039515 0.8779079 0.6917879 0.4938325 0.3026303 0.1615921
#>  [9,] 0.5514776 0.6204398 0.6805952 0.6480233 0.5898914 0.5719211 0.6446195
#> [10,] 0.8842282 0.7790874 0.6934010 0.6480343 0.4977642 0.4156656 0.3503331
#> [11,] 0.8585980 0.7665055 0.7327642 0.6219931 0.4085914 0.3332692 0.3240332
#> [12,] 0.3407420 0.2352279 0.3079892 0.4140692 0.5601267 0.7359672 0.7984598
#> [13,] 0.4290656 0.5402511 0.6672689 0.6875475 0.6638569 0.7754350 0.8103377
#> [14,] 0.4312445 0.4408813 0.3944996 0.4553784 0.4987896 0.5538183 0.7180605
#> [15,] 0.3513289 0.5546746 0.6412243 0.6837321 0.8096848 0.9070080 0.8939794
#> [16,] 0.5836060 0.6994858 0.7165754 0.6770175 0.6268818 0.6250545 0.6256866
#> [17,] 0.4721515 0.8842405 0.9671468 0.8790662 0.6717682 0.6800986 0.5572603
#> [18,] 0.2148567 0.1792948 0.2308789 0.2981014 0.2361611 0.3611904 0.5001433
#> [19,] 0.7007175 0.5574112 0.4759632 0.3483826 0.5782201 0.8728128 0.9246887
#> [20,] 0.2365486 0.4814059 0.6232290 0.5399960 0.5613107 0.7182927 0.7325890
#> [21,] 0.3387435 0.3646612 0.4419457 0.6464884 0.8497423 0.8708878 0.7547749
#> [22,] 0.2339821 0.2892487 0.4671677 0.6326240 0.5945150 0.5759544 0.5258105
#> [23,] 0.6731500 0.8916027 1.0409989 0.9161516 0.8469296 0.9227879 0.9265951
#> [24,] 0.5926924 0.7039777 0.8340989 0.8649340 0.9351957 0.9724680 0.9556030
#> [25,] 0.6403928 0.7877079 0.9378345 0.9223231 0.9150698 0.9890428 0.9924874
#>              V28       V29           V3       V30       V31       V32       V33
#>  [1,] 0.85257287 0.9726135 -0.016995815 0.8383103 0.4787207 0.4511774 0.4287571
#>  [2,] 1.06886101 0.9427509  0.025946776 0.6836506 0.6125940 0.5340099 0.3928203
#>  [3,] 1.00000000 0.7925785  0.044217644 0.5489568 0.6410281 0.6367816 0.4844584
#>  [4,] 0.85224444 0.7366744  0.047085313 0.6063815 0.4603005 0.3463824 0.2325178
#>  [5,] 0.91608364 0.7605221  0.046942324 0.5895029 0.6087893 0.6104801 0.5285812
#>  [6,] 0.35422601 0.3386680  0.057682212 0.3816265 0.3932921 0.3012344 0.3971563
#>  [7,] 0.82984464 0.7237923  0.042358102 0.5544468 0.4018085 0.3591100 0.3342970
#>  [8,] 0.09778285 0.2235528  0.014040380 0.4006485 0.3470075 0.5553766 0.4554312
#>  [9,] 0.74696359 0.7147869  0.086354329 0.7194102 0.7675255 0.7862874 0.6490926
#> [10,] 0.30142503 0.5047494  0.059509033 0.4852092 0.2557410 0.2643153 0.3644675
#> [11,] 0.35327689 0.2815513  0.083752473 0.2709404 0.1805575 0.1983911 0.2412552
#> [12,] 0.86659361 0.8374356  0.035388132 0.9309615 0.7880022 0.7054570 0.6709556
#> [13,] 0.78235671 0.7704706  0.085359651 0.5633360 0.5491780 0.3672546 0.2808431
#> [14,] 0.88866265 0.9722067  0.009325025 0.9365722 0.8339820 0.7258149 0.6925883
#> [15,] 0.81195626 0.8501018  0.036724517 0.9139645 0.8965725 0.7557328 0.6837905
#> [16,] 0.57537586 0.4711602  0.036942089 0.5367659 0.5135126 0.3998929 0.3002826
#> [17,] 0.65850605 0.8005800  0.041666679 0.7764561 0.5926552 0.6331752 0.8167260
#> [18,] 0.71558333 0.8367565  0.027748675 0.7924189 0.6640068 0.5256998 0.6635708
#> [19,] 0.49509408 0.2964517  0.058995355 0.1504136 0.1897158 0.4340355 0.4651665
#> [20,] 0.74029584 0.7610077  0.019170939 0.7582944 0.5429878 0.3445600 0.6153883
#> [21,] 0.71508412 0.6178376  0.016918806 0.4232692 0.3265781 0.2460419 0.2135160
#> [22,] 0.72598857 0.9841350  0.025353345 0.9917198 0.9425839 0.8559722 0.7495960
#> [23,] 0.89926169 0.8303020 -0.000908780 0.7234492 0.6370173 0.5310648 0.4655991
#> [24,] 0.81282939 0.7107142  0.026423137 0.6473725 0.6253512 0.5518988 0.5043302
#> [25,] 0.91945975 0.8537452  0.007741515 0.7443217 0.6752967 0.6259309 0.5690128
#>             V34         V35        V36       V37       V38        V39
#>  [1,] 0.3949496 0.284288927 0.18294000 0.1883054 0.4515481 0.46547787
#>  [2,] 0.3032557 0.131957027 0.06186766 0.2072423 0.1549486 0.18230384
#>  [3,] 0.2074331 0.096980900 0.04589166 0.1602176 0.1897663 0.18507875
#>  [4,] 0.2639072 0.199309833 0.05465886 0.1075923 0.1592990 0.13727622
#>  [5,] 0.4733621 0.279421943 0.15381367 0.2006372 0.2005945 0.31683839
#>  [6,] 0.4945917 0.466591584 0.41576992 0.3729916 0.3585554 0.41800183
#>  [7,] 0.2494975 0.152421019 0.22533882 0.2660832 0.3059291 0.36108156
#>  [8,] 0.1813934 0.005860995 0.24183676 0.4043501 0.4031342 0.30596973
#>  [9,] 0.6086347 0.605489802 0.42182566 0.2060542 0.3504125 0.39573082
#> [10,] 0.4472340 0.626939164 0.56148097 0.4822320 0.5392556 0.47699210
#> [11,] 0.2206280 0.273767419 0.16520652 0.2151739 0.3343294 0.28863121
#> [12,] 0.6852934 0.713490226 0.75607796 0.7412804 0.6795604 0.54106601
#> [13,] 0.1929772 0.071474752 0.07296379 0.2474677 0.1717184 0.25191627
#> [14,] 0.6099480 0.594918953 0.57973769 0.5158111 0.3678412 0.19662887
#> [15,] 0.7343986 0.849947476 0.92283020 0.8473791 0.6992659 0.53688165
#> [16,] 0.2000305 0.323612691 0.47530703 0.5143316 0.3325460 0.23878537
#> [17,] 0.9160078 0.875863297 0.66752218 0.3972802 0.2255490 0.25857555
#> [18,] 0.6904337 0.595030795 0.60507660 0.5901543 0.5833561 0.58155049
#> [19,] 0.3240736 0.293076587 0.36815479 0.5143606 0.3580775 0.47422333
#> [20,] 0.9472607 0.875878187 0.69512337 0.3748460 0.3306523 0.17539067
#> [21,] 0.2178516 0.266572982 0.23055948 0.1268585 0.1228440 0.14568987
#> [22,] 0.6358894 0.521738356 0.45827317 0.4303224 0.1899610 0.09135417
#> [23,] 0.5085158 0.545902829 0.48824665 0.4818057 0.3390139 0.19138372
#> [24,] 0.4959301 0.479815245 0.44124957 0.3909859 0.2842879 0.12973194
#> [25,] 0.5793938 0.577642120 0.51542569 0.4606059 0.3699056 0.18163175
#>                V4        V40        V41         V42        V43           V44
#>  [1,] 0.004001291 0.34689490 0.07758038 0.144173919 0.43571220  3.641039e-01
#>  [2,] 0.071578820 0.25650000 0.30638992 0.370024075 0.18141796  1.187928e-01
#>  [3,] 0.075191373 0.11211743 0.16302575 0.064958258 0.08297429  1.383554e-01
#>  [4,] 0.041336588 0.15300992 0.07838951 0.142739308 0.21059008  2.321663e-01
#>  [5,] 0.012704188 0.44890555 0.42545028 0.271501445 0.18291047  7.841055e-02
#>  [6,] 0.052588885 0.39719632 0.31708949 0.364437011 0.29625811  1.917185e-01
#>  [7,] 0.046752357 0.34630231 0.24835591 0.286836384 0.28614300  3.002766e-01
#>  [8,] 0.045591151 0.19531133 0.16429794 0.113832010 0.06355041  8.631507e-02
#>  [9,] 0.080940610 0.33150509 0.59020003 0.638084166 0.41754218  3.978132e-01
#> [10,] 0.080399622 0.47538233 0.37000842 0.316216818 0.42177391  3.672002e-01
#> [11,] 0.114156314 0.20611940 0.17884116 0.209535691 0.22510154  1.342742e-01
#> [12,] 0.077255905 0.48063285 0.42429879 0.349583181 0.35902343  4.398046e-01
#> [13,] 0.110596766 0.26489054 0.24687561 0.386592646 0.24344749  6.624636e-02
#> [14,] 0.005622192 0.12361892 0.10999037 0.161553288 0.21904757  2.339244e-01
#> [15,] 0.028312286 0.35961824 0.28303413 0.225576278 0.27674542  2.159034e-01
#> [16,] 0.023727268 0.35728494 0.35916648 0.297506361 0.18086662  1.465374e-01
#> [17,] 0.035291706 0.36243057 0.32005247 0.208976635 0.29835797  2.186550e-01
#> [18,] 0.032110624 0.50155638 0.34228030 0.240969604 0.17396613  1.128796e-01
#> [19,] 0.036597541 0.46037982 0.28825904 0.260382650 0.24734121  2.098909e-01
#> [20,] 0.038014482 0.36154113 0.39520411 0.448079189 0.39643886  2.073495e-01
#> [21,] 0.027656224 0.17502882 0.22012566 0.198938726 0.16763192  1.206829e-01
#> [22,] 0.041430202 0.13507733 0.15950239 0.167584743 0.26279647  3.378247e-01
#> [23,] 0.023591396 0.07978376 0.07286854 0.002418559 0.07030208  9.297447e-03
#> [24,] 0.016049779 0.19659104 0.20859300 0.123786191 0.04939557  1.544735e-02
#> [25,] 0.035595640 0.16287204 0.15254496 0.039345796 0.02465527 -2.674014e-05
#>                V45         V46        V47        V48         V49         V5
#>  [1,]  0.091053925 0.184265831 0.15597633 0.08085256 0.068520394 0.01650459
#>  [2,]  0.193120354 0.259683024 0.26556559 0.21008201 0.077113074 0.04817123
#>  [3,]  0.067580443 0.065317210 0.07670148 0.10896804 0.075004303 0.10218922
#>  [4,]  0.241289941 0.246348054 0.18939054 0.09965742 0.068281659 0.05611430
#>  [5,]  0.077400409 0.096887864 0.10919574 0.11037700 0.053432299 0.05294242
#>  [6,]  0.128133231 0.119394178 0.10173819 0.06782747 0.052256646 0.09659342
#>  [7,]  0.292206013 0.183823496 0.11272786 0.08836406 0.052741607 0.10207683
#>  [8,]  0.047854911 0.033293006 0.03446497 0.03583616 0.016669183 0.04453739
#>  [9,]  0.527903471 0.487635042 0.31067074 0.16688302 0.092884848 0.06106949
#> [10,]  0.234781521 0.190895862 0.18977917 0.11046531 0.059791907 0.07456704
#> [11,]  0.085018907 0.053432931 0.03431591 0.04813701 0.038418826 0.14570580
#> [12,]  0.424797223 0.271922465 0.15176575 0.17787242 0.092237276 0.07752642
#> [13,]  0.111496764 0.202130918 0.25238859 0.17978725 0.076513379 0.13328929
#> [14,]  0.199809111 0.137260816 0.08153234 0.09000176 0.052590148 0.01626673
#> [15,]  0.216424925 0.179611570 0.10295345 0.03317960 0.019858908 0.01712933
#> [16,]  0.147328135 0.118451746 0.12400404 0.10696587 0.062777814 0.05471000
#> [17,]  0.109940503 0.056097832 0.04157920 0.03953686 0.014264175 0.06969281
#> [18,]  0.109240841 0.062417453 0.06536778 0.06356534 0.033832580 0.03000447
#> [19,]  0.080035071 0.057040755 0.06133574 0.03728157 0.053220751 0.04832466
#> [20,]  0.101955024 0.096574644 0.11718403 0.06448388 0.023598832 0.06015891
#> [21,]  0.100004851 0.099504515 0.04905610 0.04052618 0.027005509 0.04695896
#> [22,]  0.297623165 0.193313817 0.09190431 0.02668210 0.014145460 0.03599767
#> [23,] -0.040704487 0.026204962 0.03398051 0.02593433 0.017155620 0.05756673
#> [24,]  0.027931522 0.031429035 0.01937754 0.02805503 0.010189874 0.03450349
#> [25,]  0.007520985 0.005545076 0.01024404 0.01745920 0.005756979 0.05412885
#>               V50         V51         V52          V53         V54          V55
#>  [1,] 0.027672107 0.011800988 0.018349945  0.008924824 0.002416908 0.0030037100
#>  [2,] 0.018464427 0.022618015 0.013698319  0.003970261 0.008252997 0.0077413568
#>  [3,] 0.027714798 0.015416232 0.009457997  0.005258953 0.006533898 0.0041568014
#>  [4,] 0.024338757 0.017450811 0.013359441  0.002569047 0.004485027 0.0039238065
#>  [5,] 0.006974698 0.013526885 0.015053632  0.010472536 0.002679746 0.0014163751
#>  [6,] 0.018466978 0.012630238 0.010948974  0.010546785 0.019640907 0.0118676536
#>  [7,] 0.018483471 0.017434811 0.013459970  0.012131307 0.013551097 0.0085827999
#>  [8,] 0.010283431 0.012565354 0.008056263 -0.001122120 0.005225794 0.0024482749
#>  [9,] 0.040572362 0.038384304 0.034706038  0.015730683 0.019606688 0.0229389706
#> [10,] 0.029059965 0.010135972 0.001758709  0.018615756 0.014941894 0.0003307268
#> [11,] 0.020691545 0.019531113 0.023340004  0.015879323 0.017361084 0.0199187607
#> [12,] 0.026493983 0.021591261 0.019274496  0.015219380 0.013807989 0.0073543492
#> [13,] 0.019200317 0.023371959 0.021984073  0.012525012 0.010155409 0.0076308725
#> [14,] 0.015295365 0.007786421 0.007933064  0.002805252 0.008614847 0.0074363301
#> [15,] 0.034874466 0.005158466 0.008877752  0.003574694 0.012890506 0.0100264693
#> [16,] 0.014267528 0.017007918 0.010264193  0.008597846 0.008374332 0.0068849764
#> [17,] 0.017254791 0.013559068 0.014508115  0.020257180 0.015451100 0.0066274439
#> [18,] 0.014204186 0.008139484 0.009024758  0.016763725 0.011938670 0.0106449444
#> [19,] 0.023253747 0.005260417 0.008833770  0.005559109 0.019448069 0.0196581835
#> [20,] 0.029008249 0.036913185 0.020197426  0.013459332 0.013335508 0.0062235784
#> [21,] 0.012317004 0.009190705 0.011070424  0.007735805 0.008823375 0.0084353188
#> [22,] 0.019071406 0.014116709 0.005967877  0.008797353 0.010579917 0.0019209650
#> [23,] 0.008549163 0.002163425 0.002663861  0.008434832 0.010774992 0.0103379878
#> [24,] 0.003991630 0.001720793 0.004019693  0.008043817 0.004132581 0.0060598383
#> [25,] 0.010084180 0.008125816 0.006413434  0.005989215 0.003083591 0.0016288872
#>               V56          V57         V58         V59         V6         V60
#>  [1,] 0.004757740 -0.002681338 0.004155584 0.001571190 0.03574275 0.005039918
#>  [2,] 0.003939465  0.002365570 0.007924858 0.005859922 0.04568188 0.001530806
#>  [3,] 0.009842960  0.005557518 0.006016471 0.008332942 0.11784657 0.011073273
#>  [4,] 0.003656193  0.002931810 0.003664453 0.003244679 0.06442881 0.007097011
#>  [5,] 0.004163939  0.001609434 0.003439305 0.003642574 0.06052714 0.005071753
#>  [6,] 0.006838782  0.010098111 0.013398321 0.009545213 0.08844540 0.006860265
#>  [7,] 0.008574864  0.006720946 0.009152363 0.007387052 0.15516335 0.006013604
#>  [8,] 0.004689462  0.007822344 0.005208523 0.000884399 0.07057277 0.002016775
#>  [9,] 0.010292725  0.011503863 0.011290676 0.018982143 0.13176170 0.010557358
#> [10,] 0.014651231  0.006677647 0.003418215 0.016257691 0.08617789 0.010801583
#> [11,] 0.013939543  0.007395804 0.010298276 0.017331646 0.11309857 0.010960991
#> [12,] 0.010616606  0.009160475 0.005367989 0.004156993 0.09734277 0.004240244
#> [13,] 0.003562601  0.009980111 0.010227387 0.005849994 0.15857123 0.005978192
#> [14,] 0.003492283  0.004375308 0.003745423 0.006460834 0.08530508 0.003708626
#> [15,] 0.002385958  0.010897852 0.006538310 0.007074830 0.04939692 0.008221585
#> [16,] 0.008349457  0.012321243 0.008782113 0.009093874 0.10359972 0.007402967
#> [17,] 0.013658864  0.023253846 0.009002436 0.007891330 0.06980145 0.012877277
#> [18,] 0.008746122  0.006621327 0.005411502 0.008906830 0.05575235 0.003816252
#> [19,] 0.004803555  0.011991228 0.021019022 0.014607835 0.01932713 0.004715920
#> [20,] 0.009056724  0.010049337 0.003017545 0.005184547 0.08448671 0.004594811
#> [21,] 0.004978748  0.004995382 0.004216785 0.004290731 0.06409902 0.003521387
#> [22,] 0.003204982  0.005428594 0.004418474 0.009474623 0.03337680 0.003231811
#> [23,] 0.001586097  0.003741006 0.002382608 0.002338806 0.09989499 0.002746585
#> [24,] 0.006393477  0.002322579 0.004245773 0.009119965 0.07201499 0.007141422
#> [25,] 0.004465659  0.004141998 0.002886417 0.001945900 0.08404416 0.001803936
#>               V7          V8         V9
#>  [1,] 0.05311122 -0.02028804 0.07652973
#>  [2,] 0.10518579  0.05518201 0.10314598
#>  [3,] 0.14462319  0.22449233 0.30005891
#>  [4,] 0.11237254  0.14165508 0.22054932
#>  [5,] 0.04258974  0.04089399 0.18040452
#>  [6,] 0.09100693  0.11172148 0.17398934
#>  [7,] 0.14641012  0.19417069 0.30230999
#>  [8,] 0.14757158  0.14393147 0.22074133
#>  [9,] 0.09102152  0.13190925 0.21190738
#> [10,] 0.05884109  0.14874644 0.10774636
#> [11,] 0.12794721  0.12824176 0.15095031
#> [12,] 0.13609338  0.20918432 0.39639068
#> [13,] 0.13582721  0.13359164 0.09484226
#> [14,] 0.08333258  0.03640309 0.06786538
#> [15,] 0.12552281  0.13578332 0.07197129
#> [16,] 0.14795583  0.22362843 0.25611876
#> [17,] 0.10374600  0.08648760 0.02957343
#> [18,] 0.04252519  0.06526236 0.05270904
#> [19,] 0.02940924  0.00151994 0.06372066
#> [20,] 0.11903696  0.13098727 0.12772559
#> [21,] 0.07238336  0.07174988 0.09772334
#> [22,] 0.07893171  0.10073005 0.10739859
#> [23,] 0.10510773  0.08112602 0.14723388
#> [24,] 0.09246767  0.07651424 0.12505747
#> [25,] 0.08694972  0.08836704 0.16410934
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551