Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1        V10        V11        V12        V13        V14
#>  [1,] 0.023474489 0.37689026 0.46460229 0.50629496 0.49494077 0.39263260
#>  [2,] 0.070700000 0.12470000 0.23400000 0.17640000 0.22840000 0.31150000
#>  [3,] 0.049914107 0.22666484 0.33947557 0.43400574 0.49008714 0.54185534
#>  [4,] 0.018795768 0.32147087 0.29494125 0.28496788 0.25548295 0.18058557
#>  [5,] 0.008730469 0.19784289 0.19938733 0.13569012 0.02743883 0.11964833
#>  [6,] 0.016942634 0.36194637 0.48852164 0.50470204 0.42316558 0.29436318
#>  [7,] 0.053379289 0.17525707 0.28067377 0.31020545 0.36827309 0.31993789
#>  [8,] 0.052467406 0.10970120 0.15007717 0.11036783 0.15278796 0.21987166
#>  [9,] 0.033729387 0.17856750 0.21371481 0.27399445 0.32098241 0.31603298
#> [10,] 0.021301439 0.05318815 0.12549906 0.21965613 0.27351835 0.30012672
#> [11,] 0.031358113 0.18737182 0.26878070 0.28540177 0.24682115 0.22681757
#> [12,] 0.043000000 0.71060000 0.73420000 0.50330000 0.30000000 0.19510000
#> [13,] 0.026110369 0.27337275 0.32047869 0.32959215 0.25701731 0.18125196
#> [14,] 0.006867573 0.11412685 0.07934412 0.20218625 0.01711123 0.22749020
#> [15,] 0.015352260 0.17320907 0.15264477 0.11803619 0.10651895 0.13239225
#> [16,] 0.017525135 0.17981746 0.12693842 0.13960440 0.21755136 0.21485233
#> [17,] 0.020100000 0.10970000 0.08410000 0.09420000 0.12040000 0.04200000
#> [18,] 0.013200323 0.08585076 0.10826241 0.13771593 0.22589299 0.26460579
#> [19,] 0.021540544 0.07974941 0.06295116 0.17068743 0.15433483 0.11053330
#> [20,] 0.026047698 0.10949137 0.15000467 0.15081369 0.22885014 0.37240782
#> [21,] 0.011854198 0.09212899 0.11623181 0.11869641 0.13543591 0.19935767
#> [22,] 0.006198409 0.18097691 0.19864415 0.16256823 0.21688394 0.11340411
#> [23,] 0.029800000 0.08520000 0.24760000 0.36450000 0.27770000 0.28260000
#> [24,] 0.047414003 0.06696530 0.20459447 0.41037374 0.41251846 0.13683565
#> [25,] 0.018652665 0.09927461 0.09188565 0.08399868 0.09307103 0.09616236
#>             V15       V16       V17       V18       V19          V2       V20
#>  [1,] 0.2830554 0.2903337 0.2821303 0.2840242 0.4402565 0.036540067 0.6935679
#>  [2,] 0.4725000 0.5543000 0.5386000 0.3746000 0.4583000 0.125200000 0.5961000
#>  [3,] 0.6192644 0.7343019 0.8339561 0.8802372 0.9285279 0.054197263 0.9277646
#>  [4,] 0.1362410 0.2395894 0.2679082 0.2991142 0.2391141 0.024395745 0.1661917
#>  [5,] 0.2306115 0.5182858 0.6847527 0.7780939 0.7828493 0.011639385 0.6739853
#>  [6,] 0.2103058 0.2956486 0.3945174 0.4354249 0.4795660 0.015691157 0.5661665
#>  [7,] 0.2454383 0.2578334 0.3102676 0.2133459 0.2926117 0.075431948 0.4230077
#>  [8,] 0.4036706 0.5439876 0.7118457 0.7770235 0.8690966 0.040614149 0.9670799
#>  [9,] 0.4064132 0.5003468 0.5375745 0.6051192 0.6667043 0.042159720 0.7305131
#> [10,] 0.3466989 0.4723795 0.5912647 0.6811078 0.8765263 0.010272713 1.0011670
#> [11,] 0.1904363 0.1848906 0.2628402 0.3162851 0.3892204 0.053304283 0.3897072
#> [12,] 0.2767000 0.3737000 0.2507000 0.2507000 0.3292000 0.090200000 0.4871000
#> [13,] 0.1674824 0.1939507 0.1968767 0.2828294 0.4288381 0.038420254 0.5250541
#> [14,] 0.1664025 0.2087116 0.0622953 0.2276615 0.4060000 0.014140602 0.3841080
#> [15,] 0.1444906 0.1815864 0.2329563 0.2615956 0.3356028 0.008690856 0.5554702
#> [16,] 0.2572508 0.3752537 0.4348446 0.4045688 0.4757072 0.015580500 0.4481454
#> [17,] 0.0031000 0.0162000 0.0624000 0.2127000 0.3436000 0.002600000 0.3813000
#> [18,] 0.2611299 0.2431130 0.2115607 0.2753005 0.3731278 0.017082199 0.4749568
#> [19,] 0.1196125 0.1939187 0.1763008 0.1977771 0.2050849 0.027455158 0.1876002
#> [20,] 0.5887937 0.7534145 0.8432143 0.8341384 0.8349992 0.026557159 0.8179700
#> [21,] 0.1882314 0.1969878 0.2067643 0.2262878 0.2453305 0.015518516 0.2687329
#> [22,] 0.2308468 0.5011874 0.7415241 0.7619801 0.5519629 0.013022964 0.3441488
#> [23,] 0.3237000 0.4335000 0.5638000 0.4555000 0.4348000 0.061500000 0.6433000
#> [24,] 0.2277700 0.3582229 0.5735916 0.6673121 0.5465245 0.049018553 0.2805336
#> [25,] 0.0939552 0.1895738 0.3116317 0.2923130 0.2492358 0.021594325 0.3366134
#>             V21       V22       V23       V24       V25       V26       V27
#>  [1,] 0.8079684 0.7860728 0.6750197 0.7071499 0.7133835 0.8508329 0.9702917
#>  [2,] 0.7464000 0.7644000 0.5711000 0.6257000 0.6695000 0.7131000 0.7567000
#>  [3,] 0.9341849 0.8021229 0.6893820 0.6418599 0.5077301 0.3607481 0.2971534
#>  [4,] 0.1124786 0.1636615 0.3086404 0.4715573 0.6304257 0.7868701 0.8864588
#>  [5,] 0.7028866 0.8377237 0.8863954 0.4828397 0.1240704 0.2828554 0.6256856
#>  [6,] 0.7466867 0.9303171 0.9745662 0.9725884 0.9919957 0.9046352 0.8132803
#>  [7,] 0.5353379 0.6138049 0.6662094 0.6337498 0.5727717 0.5558315 0.6033909
#>  [8,] 1.0043956 1.0325618 0.9369108 0.7368534 0.5415643 0.4512747 0.3483298
#>  [9,] 0.7465830 0.8164752 0.9294652 0.9731541 0.9612663 1.0061807 0.9867605
#> [10,] 0.9896212 1.0210630 1.0441041 0.9596734 0.9285622 0.8525572 0.7914895
#> [11,] 0.4333149 0.4613700 0.5313281 0.6023949 0.6249379 0.6917864 0.8137833
#> [12,] 0.6527000 0.8454000 0.9739000 1.0000000 0.6665000 0.5323000 0.4024000
#> [13,] 0.6420510 0.7337594 0.7451477 0.8309564 0.9007076 0.9856904 1.0018115
#> [14,] 0.2531098 0.3660466 0.5703552 0.4926638 0.3193460 0.5407985 0.5364561
#> [15,] 0.6730922 0.8474699 0.9852380 0.9142777 0.8888732 0.9388265 0.9113034
#> [16,] 0.5182370 0.4933389 0.4999471 0.5767310 0.7279782 0.8414339 0.8046079
#> [17,] 0.3825000 0.4764000 0.6313000 0.7523000 0.8675000 0.8788000 0.7901000
#> [18,] 0.5351257 0.5890051 0.5799395 0.6566815 0.8203932 0.8906761 0.9129429
#> [19,] 0.1957921 0.1473126 0.2558827 0.2609627 0.1044982 0.2588792 0.5688976
#> [20,] 0.8840977 0.9065483 0.8370688 0.8441605 0.8435997 0.8792457 0.8017951
#> [21,] 0.3064212 0.4018811 0.5013194 0.5958610 0.6407893 0.6875991 0.6862882
#> [22,] 0.4844535 0.3593815 0.3568581 0.6456609 0.8052150 0.7503773 0.7043524
#> [23,] 0.3932000 0.1989000 0.3540000 0.9165000 0.9371000 0.4620000 0.2771000
#> [24,] 0.1065826 0.1311931 0.2290324 0.1969244 0.2586316 0.2439232 0.2752050
#> [25,] 0.4945530 0.6935503 0.8500875 0.8460754 0.8167882 0.8834692 0.9337995
#>             V28       V29           V3       V30       V31       V32       V33
#>  [1,] 0.9681821 0.8546100  0.055367196 0.5983887 0.4929760 0.5609805 0.4777750
#>  [2,] 0.8077000 0.8477000  0.144700000 0.9289000 0.9513000 0.7995000 0.4362000
#>  [3,] 0.2012911 0.1859256  0.058151648 0.2140989 0.1373003 0.1299245 0.2682972
#>  [4,] 0.9041654 0.9053176  0.021612593 0.9712265 0.9434666 0.8216753 0.6921963
#>  [5,] 0.8580841 0.9485511 -0.010754237 0.8063565 0.4947232 0.5095953 0.4509078
#>  [6,] 0.6721369 0.4159070  0.025188530 0.3320392 0.3078322 0.4497521 0.4977665
#>  [7,] 0.7458449 0.7536057  0.072661706 0.7695030 0.8468073 0.7976994 0.6285068
#>  [8,] 0.3806250 0.4190930  0.039178375 0.4753832 0.3611929 0.4094237 0.4251460
#>  [9,] 0.9598399 0.7921873  0.014642279 0.5862708 0.5103879 0.4610039 0.3055329
#> [10,] 0.7369680 0.5918288  0.034942404 0.4203325 0.2138085 0.1422320 0.1599763
#> [11,] 0.8311130 0.7713596  0.068210444 0.6323779 0.5347502 0.4520156 0.4131957
#> [12,] 0.3444000 0.4239000  0.083300000 0.4182000 0.4393000 0.1162000 0.4336000
#> [13,] 0.9363216 0.8330451  0.039647406 0.6855863 0.5484578 0.3799194 0.2995689
#> [14,] 0.2516152 0.2016164  0.061575067 0.3471203 0.6361789 0.7460643 0.6142374
#> [15,] 0.8507350 0.7782069  0.008050501 0.6934990 0.6581718 0.5684326 0.5246727
#> [16,] 0.5638950 0.3044015  0.032325501 0.2460196 0.3710613 0.5177054 0.4867386
#> [17,] 0.8357000 0.9631000  0.013800000 0.9619000 0.9236000 0.8903000 0.9708000
#> [18,] 0.8436203 0.7110025  0.014119045 0.5575515 0.4992259 0.3852196 0.2837725
#> [19,] 0.7854327 0.8295414  0.019421592 0.7690572 0.6932685 0.4530176 0.6059894
#> [20,] 0.6352195 0.5248000  0.033011135 0.5273056 0.4889562 0.3198456 0.3135114
#> [21,] 0.7939265 0.8983948  0.018298756 0.8876625 0.7562947 0.6254397 0.6416386
#> [22,] 0.8292006 0.6385283  0.019360657 0.3765508 0.3604921 0.3746287 0.2903395
#> [23,] 0.6613000 0.8028000  0.065000000 0.4200000 0.5192000 0.6962000 0.5792000
#> [24,] 0.4827864 0.7191631  0.080987403 0.7084538 0.8243290 0.8657547 0.7108603
#> [25,] 0.9058993 0.9248852  0.023724901 0.8478212 0.7938582 0.7296676 0.6816644
#>             V34        V35        V36       V37       V38        V39
#>  [1,] 0.2897406 0.06661347 0.26166024 0.2554186 0.3706125 0.59006706
#>  [2,] 0.4048000 0.49520000 0.17120000 0.3652000 0.3763000 0.28410000
#>  [3,] 0.3575011 0.37481156 0.30156396 0.2532208 0.2811653 0.25169175
#>  [4,] 0.6313319 0.67169126 0.73452158 0.7348452 0.6360035 0.53856122
#>  [5,] 0.3834777 0.27592492 0.20421184 0.2320616 0.4738557 0.47433824
#>  [6,] 0.4303301 0.24766445 0.28896438 0.3527700 0.3149768 0.46140866
#>  [7,] 0.6417254 0.66193052 0.43007463 0.1954994 0.3006019 0.34435172
#>  [8,] 0.3444313 0.23420835 0.26627271 0.3217150 0.3758421 0.35852530
#>  [9,] 0.1364695 0.06402058 0.09257354 0.1992320 0.2877731 0.29147309
#> [10,] 0.2465105 0.33743975 0.35224340 0.2838310 0.2635999 0.40087968
#> [11,] 0.3218131 0.19237494 0.18714393 0.2589509 0.2000805 0.20335440
#> [12,] 0.6553000 0.61720000 0.43730000 0.4118000 0.3641000 0.45720000
#> [13,] 0.2161230 0.13703086 0.05049443 0.1070357 0.1415634 0.13962556
#> [14,] 0.3403624 0.38135921 0.28468763 0.5335208 0.8807974 0.98877276
#> [15,] 0.5384143 0.53767900 0.47058766 0.4471333 0.3131201 0.14703953
#> [16,] 0.4970993 0.55528809 0.56864713 0.5032199 0.3580291 0.39544138
#> [17,] 0.9647000 0.78920000 0.53070000 0.2718000 0.1953000 0.13740000
#> [18,] 0.2236289 0.21282002 0.24699411 0.2327485 0.2551556 0.24376519
#> [19,] 0.7542851 0.70119585 0.71186238 0.6211059 0.5600082 0.52511659
#> [20,] 0.3104423 0.21751350 0.13784059 0.2460942 0.1779691 0.08335264
#> [21,] 0.6390515 0.59990798 0.51310384 0.3964372 0.3026066 0.24622232
#> [22,] 0.1631667 0.27798982 0.29786631 0.1592580 0.1160113 0.12437548
#> [23,] 0.8889000 0.78630000 0.71330000 0.7615000 0.4401000 0.30090000
#> [24,] 0.6572393 0.68535290 0.81917045 0.9303518 1.0255526 0.92068067
#> [25,] 0.7082391 0.71491723 0.63734968 0.5339075 0.4264130 0.32104808
#>                V4       V40        V41        V42        V43        V44
#>  [1,] 0.042048842 0.4709782 0.25579222 0.45263318 0.46579558 0.49365935
#>  [2,] 0.164400000 0.0427000 0.53310000 0.69520000 0.42880000 0.30630000
#>  [3,] 0.081613721 0.1926941 0.27431994 0.30234731 0.23910459 0.17621913
#>  [4,] 0.053653059 0.5362073 0.46855333 0.38713711 0.42630875 0.52003247
#>  [5,] 0.006025688 0.3445938 0.08228305 0.17439072 0.45819953 0.36661259
#>  [6,] 0.032788687 0.4719551 0.29625363 0.09291004 0.30548084 0.33168088
#>  [7,] 0.067711554 0.3639914 0.62907490 0.67488612 0.46800422 0.40842519
#>  [8,] 0.030755710 0.2204745 0.20369791 0.27091078 0.18720642 0.22037995
#>  [9,] 0.029839171 0.2725868 0.26969690 0.24203573 0.18873131 0.18584418
#> [10,] 0.114822897 0.3949919 0.39306991 0.41872923 0.32497157 0.22667122
#> [11,] 0.096848593 0.2751664 0.32357710 0.37367179 0.29443015 0.16642417
#> [12,] 0.081300000 0.4367000 0.29640000 0.43120000 0.41550000 0.18240000
#> [13,] 0.049545703 0.1164221 0.06549829 0.12631713 0.19292934 0.19583236
#> [14,] 0.020043561 0.9258914 0.62197759 0.51283138 0.32706498 0.32171550
#> [15,] 0.031773647 0.1173133 0.12138548 0.06867378 0.09367615 0.01638979
#> [16,] 0.028501021 0.4644313 0.38203861 0.31629366 0.23137903 0.19424593
#> [17,] 0.006200000 0.3105000 0.37900000 0.41050000 0.33550000 0.29980000
#> [18,] 0.018283772 0.1822120 0.15706065 0.19361662 0.17621717 0.12997070
#> [19,] 0.025875180 0.5816889 0.48750126 0.42304404 0.28278749 0.13995256
#> [20,] 0.036140649 0.1462116 0.10799056 0.06953224 0.10470380 0.11723153
#> [21,] 0.028628201 0.2136460 0.20332927 0.17414989 0.21543078 0.18851490
#> [22,] 0.035700392 0.1213365 0.09976992 0.09308343 0.10068666 0.11040420
#> [23,] 0.092100000 0.3163000 0.28090000 0.28980000 0.05260000 0.18670000
#> [24,] 0.009645574 0.7859277 0.80244976 0.69224711 0.50941390 0.34909995
#> [25,] 0.036144498 0.2613229 0.18776840 0.11923727 0.11930862 0.09123733
#>               V45        V46        V47        V48         V49         V5
#>  [1,]  0.49883840 0.21826822 0.14144221 0.18159991 0.102287120 0.09082889
#>  [2,]  0.58350000 0.56920000 0.26300000 0.11960000 0.098300000 0.16930000
#>  [3,]  0.14702612 0.04856060 0.05885067 0.04304839 0.028429100 0.10440453
#>  [4,]  0.45706304 0.28665796 0.17288031 0.17609699 0.099124083 0.08418760
#>  [5,]  0.08420363 0.19080227 0.15552659 0.07849709 0.069583244 0.01604174
#>  [6,]  0.34085375 0.27836398 0.16300499 0.08351134 0.050379772 0.04049951
#>  [7,]  0.56327794 0.54370830 0.32861550 0.17443821 0.088119531 0.06911434
#>  [8,]  0.16636139 0.08561555 0.05010198 0.02487637 0.041894827 0.08512805
#>  [9,]  0.13861706 0.10419858 0.08930458 0.05729465 0.027435718 0.06638557
#> [10,]  0.17634590 0.12564061 0.11530524 0.05540858 0.027464542 0.20042856
#> [11,]  0.18526008 0.16201402 0.16922152 0.15126235 0.079313091 0.09803941
#> [12,]  0.14870000 0.01380000 0.11640000 0.20520000 0.106900000 0.01650000
#> [13,]  0.17141294 0.16183113 0.13754776 0.09332284 0.050916800 0.05953061
#> [14,]  0.42675838 0.36367567 0.26643973 0.15654989 0.066873878 0.02041945
#> [15,] -0.01874013 0.02297032 0.03281088 0.03225977 0.015746066 0.06743639
#> [16,]  0.02713768 0.01688872 0.02891443 0.03206933 0.018471305 0.04327686
#> [17,]  0.27480000 0.20240000 0.10430000 0.04530000 0.033700000 0.01330000
#> [18,]  0.11506207 0.09960308 0.09007154 0.06361954 0.033765893 0.03700213
#> [19,]  0.09140155 0.05489044 0.08670959 0.06917318 0.032517353 0.04864929
#> [20,]  0.12491465 0.07908369 0.09595282 0.07632178 0.044813542 0.04571311
#> [21,]  0.13652277 0.08398531 0.05378269 0.03718633 0.019519894 0.03253174
#> [22,]  0.08755204 0.08330252 0.03206574 0.04079217 0.010890233 0.05889249
#> [23,]  0.15530000 0.16330000 0.12520000 0.07480000 0.045200000 0.16150000
#> [24,]  0.13096190 0.15206389 0.20681878 0.15039953 0.083930257 0.03431177
#> [25,]  0.03662950 0.05270898 0.03848163 0.03418433 0.009254603 0.06063844
#>               V50         V51         V52          V53          V54         V55
#>  [1,] 0.020428244 0.032977649 0.019294958  0.007040895 0.0107952983 0.011998891
#>  [2,] 0.037400000 0.029100000 0.015600000  0.019700000 0.0135000000 0.012700000
#>  [3,] 0.015547494 0.021879930 0.014860341  0.017258532 0.0159650065 0.016881182
#>  [4,] 0.025847516 0.018489333 0.017456113  0.012445588 0.0079278894 0.005274751
#>  [5,] 0.025644000 0.011885029 0.018651895  0.009480176 0.0021086355 0.002665457
#>  [6,] 0.022564749 0.012092161 0.005520175  0.010348944 0.0141262412 0.013795411
#>  [7,] 0.041537807 0.039376524 0.035605258  0.019689903 0.0183701298 0.019861047
#>  [8,] 0.013536404 0.004220035 0.001364253 -0.001975020 0.0155155929 0.011214595
#>  [9,] 0.008101691 0.015838116 0.009715987  0.013854846 0.0128605405 0.010165172
#> [10,] 0.020551755 0.017964000 0.017275013  0.017128920 0.0135757997 0.004039077
#> [11,] 0.015083824 0.020092825 0.017949426  0.012065248 0.0113713306 0.008320815
#> [12,] 0.019900000 0.020800000 0.017600000  0.019700000 0.0210000000 0.014100000
#> [13,] 0.016301118 0.015949488 0.010826099  0.005722794 0.0062470981 0.003307756
#> [14,] 0.028764565 0.024138782 0.012165632  0.003513485 0.0151939122 0.008258356
#> [15,] 0.008830826 0.004440052 0.003094074  0.008371716 0.0091122698 0.009032554
#> [16,] 0.012072677 0.012567376 0.004343676  0.005801972 0.0130723045 0.008206253
#> [17,] 0.012200000 0.007200000 0.010800000  0.007000000 0.0063000000 0.003000000
#> [18,] 0.018592426 0.008058983 0.007960635  0.007853664 0.0082064276 0.005311529
#> [19,] 0.016019454 0.010598757 0.012249047  0.020904618 0.0164948315 0.010185243
#> [20,] 0.022470376 0.015039038 0.005788598  0.008294681 0.0083239990 0.006661798
#> [21,] 0.012897937 0.010751062 0.007963907  0.009268375 0.0087320298 0.005992376
#> [22,] 0.016818250 0.005877596 0.018882560  0.009779089 0.0141984432 0.014110688
#> [23,] 0.006400000 0.015400000 0.003100000  0.015300000 0.0071000000 0.021200000
#> [24,] 0.030029138 0.017956127 0.011970178  0.014790405 0.0007090551 0.003941907
#> [25,] 0.012945857 0.004228966 0.005798593  0.003951142 0.0057640904 0.005833483
#>                V56           V57         V58         V59         V6         V60
#>  [1,] 0.0032255082  0.0054710097 0.003316789 0.004490391 0.12604579 0.003484433
#>  [2,] 0.0138000000  0.0133000000 0.013100000 0.015400000 0.08440000 0.021800000
#>  [3,] 0.0119161805  0.0117013879 0.012955595 0.018136347 0.13346018 0.013418371
#>  [4,] 0.0056052737  0.0058798497 0.004887319 0.002530346 0.11277445 0.002692883
#>  [5,] 0.0048608769 -0.0013699874 0.005042765 0.001188795 0.02451494 0.003786878
#>  [6,] 0.0121762039  0.0089645400 0.011754154 0.010682261 0.07757410 0.006666177
#>  [7,] 0.0108831517  0.0084906237 0.012365052 0.016663144 0.17069167 0.010633489
#>  [8,] 0.0086073592  0.0098105393 0.011018731 0.007496288 0.07467340 0.006497207
#>  [9,] 0.0106307351  0.0069234521 0.004047612 0.009038827 0.10024764 0.005071215
#> [10,] 0.0112685042  0.0107019827 0.011491539 0.015795120 0.08150524 0.012091165
#> [11,] 0.0062015794  0.0081273212 0.009483108 0.008116671 0.09679888 0.005615552
#> [12,] 0.0049000000  0.0027000000 0.016200000 0.005900000 0.02770000 0.002100000
#> [13,] 0.0035970290  0.0042321835 0.003741122 0.003968297 0.08320726 0.006007807
#> [14,] 0.0071896192  0.0045823430 0.004110623 0.003916469 0.03411805 0.011792481
#> [15,] 0.0034578281  0.0046406969 0.003730995 0.002726650 0.10956840 0.004055195
#> [16,] 0.0100666593  0.0132483063 0.006929754 0.008798445 0.09796361 0.003499213
#> [17,] 0.0011000000  0.0007000000 0.002400000 0.005700000 0.01510000 0.004400000
#> [18,] 0.0057349833  0.0058198845 0.006915526 0.002675730 0.04970878 0.003533477
#> [19,] 0.0079124804  0.0078178445 0.005964321 0.009373763 0.07060849 0.008862064
#> [20,] 0.0075492926  0.0043538108 0.002989846 0.004707287 0.07841502 0.005161811
#> [21,] 0.0059618267  0.0076031860 0.004435238 0.007224302 0.05095426 0.005131796
#> [22,] 0.0082506947  0.0076441055 0.004707591 0.006188200 0.06355632 0.004247623
#> [23,] 0.0076000000  0.0152000000 0.004900000 0.020000000 0.22940000 0.007300000
#> [24,] 0.0002832387 -0.0008661422 0.018070429 0.012771085 0.09221726 0.008721328
#> [25,] 0.0050873797  0.0090977895 0.007397641 0.003500975 0.08553601 0.004140410
#>               V7          V8         V9
#>  [1,] 0.11922816  0.18268590 0.26767474
#>  [2,] 0.07150000  0.09470000 0.15830000
#>  [3,] 0.15485222  0.10473698 0.10283352
#>  [4,] 0.13667495  0.16850653 0.29817302
#>  [5,] 0.03875993 -0.01205583 0.10591117
#>  [6,] 0.09092626  0.15737509 0.30149824
#>  [7,] 0.12633252  0.16119892 0.22373816
#>  [8,] 0.10187501  0.08734329 0.19009361
#>  [9,] 0.11394841  0.13457452 0.15616385
#> [10,] 0.07321440  0.10876937 0.05013575
#> [11,] 0.10257365  0.09461576 0.14583269
#> [12,] 0.05690000  0.20570000 0.38870000
#> [13,] 0.10332774  0.14489938 0.20736180
#> [14,] 0.10586507  0.12346818 0.04861873
#> [15,] 0.11151602  0.10613251 0.15526974
#> [16,] 0.11006333  0.08808659 0.12944042
#> [17,] 0.05410000  0.02100000 0.05050000
#> [18,] 0.07033416  0.06202339 0.08704649
#> [19,] 0.04601647  0.07031807 0.07684449
#> [20,] 0.13157339  0.08686894 0.10575955
#> [21,] 0.05729840  0.05970665 0.07516896
#> [22,] 0.04997308  0.04414458 0.08362805
#> [23,] 0.21760000  0.20330000 0.14590000
#> [24,] 0.10465223  0.10218317 0.10226487
#> [25,] 0.08128808  0.11165527 0.10572494
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M R R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696