Classification Learning Vector Quantization 1

Learning Vector Quantization 1. Calls class::lvqinit(), class::lvq1(), and class::lvqtest() from class.

Note

The learner does not work on tasks with two target groups and one feature.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.lvq1")

Meta Information

Task type: “classif”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, class

Parameters

Id	Type	Default	Range

alpha	numeric	0.03	\([0, \infty)\)

k	integer	5	\([1, \infty)\)

niter	integer	NULL	\([1, \infty)\)

prior	untyped	NULL	-

size	integer	NULL	\([1, \infty)\)

References

Kohonen, Teuvo (1990). “The self-organizing map.” Proceedings of the IEEE, 78(9), 1464–1480. doi:10.1109/5.58325 .

Kohonen, Teuvo (1995). Self-Organizing Maps, volume 30 series Springer Series in Information Sciences. Springer, Berlin. ISBN 978-3540967120.

See also

Author

awinterstetter

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLvq1

Methods

Public methods

LearnerClassifLvq1$new()

LearnerClassifLvq1$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLvq1$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLvq1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.lvq1")
print(learner)
#> 
#> ── <LearnerClassifLvq1> (classif.lvq1): Learning Vector Quantization 1 ─────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and class
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $x
#>                V1         V10         V11        V12        V13       V14
#>  [1,] 0.024313720  0.26761667  0.30703296 0.33338273 0.30239763 0.2361333
#>  [2,] 0.102100000  0.13690000  0.25090000 0.26310000 0.27960000 0.2977000
#>  [3,] 0.055160575  0.23774363  0.34250312 0.44623389 0.49787622 0.4466940
#>  [4,] 0.008543142  0.20653715  0.30354312 0.33159254 0.30987913 0.1970926
#>  [5,] 0.017188625  0.31553921  0.27858586 0.29382005 0.24299146 0.1777167
#>  [6,] 0.027599701  0.22898604  0.28754740 0.32298802 0.35252768 0.3582975
#>  [7,] 0.025153525  0.20611973  0.26015435 0.23136538 0.38993556 0.6019400
#>  [8,] 0.054452361 -0.18594468 -0.01160627 0.14373041 0.17230818 0.2322279
#>  [9,] 0.047899046  0.20563947  0.22494234 0.27676130 0.30072131 0.2847690
#> [10,] 0.069599096  0.42300922  0.38781791 0.21813478 0.19027983 0.2680535
#> [11,] 0.037335418  0.70535976  0.73976316 0.51200027 0.32703502 0.2118901
#> [12,] 0.027959537  0.19224443  0.27464091 0.30244792 0.28416964 0.2701659
#> [13,] 0.050393226  0.28865371  0.38392959 0.56281714 0.61832358 0.7940800
#> [14,] 0.045033468  0.05104881  0.21825588 0.39885013 0.43241490 0.5277311
#> [15,] 0.012802526  0.10283980  0.08460447 0.17042840 0.05044301 0.2058881
#> [16,] 0.023687348  0.16909451  0.22525771 0.18766535 0.29149017 0.4070917
#> [17,] 0.006248306  0.21193613  0.20918235 0.10702880 0.18927145 0.1604519
#> [18,] 0.025863689  0.08147617  0.23668825 0.36524091 0.28756412 0.2779452
#> [19,] 0.008211712  0.08320169  0.10502989 0.13354241 0.18297788 0.2171775
#> [20,] 0.023705863  0.08714668  0.07596493 0.18265720 0.15476547 0.1140454
#> [21,] 0.026938713  0.30169092  0.29399441 0.37383671 0.39189352 0.2690123
#> [22,] 0.009400000  0.13280000  0.13030000 0.02730000 0.06440000 0.0712000
#> [23,] 0.019772441  0.11542038  0.12048854 0.06683598 0.10548427 0.1301183
#> [24,] 0.018058948  0.09041413  0.08315040 0.06363908 0.06736534 0.1021419
#> [25,] 0.025689333  0.13244488  0.12380061 0.13552028 0.20923976 0.3658369
#>              V15       V16       V17       V18       V19           V2       V20
#>  [1,] 0.18809743 0.2186298 0.2235609 0.3189867 0.4932576  0.034072976 0.5977571
#>  [2,] 0.38230000 0.3129000 0.3956000 0.2093000 0.3218000  0.083000000 0.3345000
#>  [3,] 0.52834112 0.6863541 0.8747925 0.9140368 0.9281106  0.112576344 0.8666250
#>  [4,] 0.05536175 0.2126780 0.4115742 0.5613211 0.6806104 -0.004301950 0.6891958
#>  [5,] 0.14784705 0.1966847 0.2031902 0.2289205 0.1819576  0.024556057 0.1246606
#>  [6,] 0.37338397 0.4621125 0.5348194 0.6068529 0.7008393  0.028073389 0.7723919
#>  [7,] 0.68083968 0.7490887 0.6902397 0.8145343 0.9525525  0.035404348 0.9998905
#>  [8,] 0.34138641 0.4424617 0.6444939 0.7370685 0.9404389  0.040832357 1.0149244
#>  [9,] 0.29705537 0.3617739 0.3317010 0.2037297 0.3113743  0.079554705 0.4397690
#> [10,] 0.30982073 0.2744576 0.1595661 0.2292451 0.4562827  0.086678503 0.8710087
#> [11,] 0.24386670 0.2928441 0.2189508 0.2797360 0.4107637  0.072955354 0.5572020
#> [12,] 0.24860160 0.2432917 0.2604761 0.2994154 0.3678600  0.046772349 0.4164636
#> [13,] 0.82538367 0.8930230 1.0243101 0.8961794 0.9157951  0.046599105 0.8443611
#> [14,] 0.65887794 0.6583250 0.8087231 0.8815589 0.9833329  0.042059780 0.9873649
#> [15,] 0.17240492 0.2276199 0.1007574 0.2407253 0.4337995  0.023287895 0.3445544
#> [16,] 0.52361227 0.6182087 0.5192439 0.4823197 0.5206836  0.052580971 0.5341444
#> [17,] 0.30259628 0.6043080 0.7894512 0.8196652 0.6263407  0.016761157 0.3615277
#> [18,] 0.30313894 0.4360453 0.5855689 0.4929905 0.4557704  0.056589287 0.6149099
#> [19,] 0.20192652 0.1998159 0.2110028 0.2487190 0.2874969  0.008488765 0.3518824
#> [20,] 0.14395283 0.1576973 0.1487212 0.1692161 0.1920324  0.035283033 0.1710909
#> [21,] 0.29024109 0.4683444 0.7517456 0.6104567 0.2126967  0.033075532 0.2734313
#> [22,] 0.12040000 0.0717000 0.1224000 0.2349000 0.3684000  0.016600000 0.3918000
#> [23,] 0.13371355 0.1762860 0.2175072 0.2306359 0.2585910  0.010332986 0.3508911
#> [24,] 0.04437623 0.1395566 0.1646685 0.3635027 0.3529838  0.019694396 0.3196260
#> [25,] 0.54401860 0.7065423 0.7734749 0.7696791 0.8204710  0.026780739 0.8627312
#>             V21       V22       V23       V24        V25       V26       V27
#>  [1,] 0.7103036 0.7511735 0.7834647 0.8836678 0.92668197 0.9811798 0.9979420
#>  [2,] 0.3184000 0.2887000 0.3610000 0.2566000 0.41060000 0.4591000 0.4722000
#>  [3,] 0.8320878 0.5751236 0.3504625 0.3684183 0.25466827 0.1028830 0.1021712
#>  [4,] 0.7543505 0.8242304 0.9278541 0.9630704 1.02446905 0.8936725 0.5246749
#>  [5,] 0.1014432 0.2275498 0.4237166 0.5793514 0.70349183 0.8262207 0.9132618
#>  [6,] 0.8248215 0.8981291 0.9300056 0.9327571 0.91111506 0.9124983 0.9030193
#>  [7,] 1.0141294 1.0088605 0.9907189 0.8492571 0.65557609 0.6432053 0.6596239
#>  [8,] 0.8796773 0.7990851 0.8323794 0.8149799 0.60523297 0.4360523 0.4517427
#>  [9,] 0.5468333 0.5421044 0.5245890 0.5896017 0.55071470 0.5772065 0.7194992
#> [10,] 0.8492104 0.3964123 0.1153012 0.1727836 0.33443797 0.7087336 0.6904479
#> [11,] 0.6914847 0.8489001 0.9512366 1.0038603 0.72723806 0.5998883 0.5087793
#> [12,] 0.4863446 0.5449928 0.6035130 0.6705289 0.69860923 0.7400717 0.8276183
#> [13,] 0.8171880 0.8326792 0.6380970 0.3126427 0.19025063 0.1169804 0.2652146
#> [14,] 0.9151367 0.8740147 0.8828174 0.8100008 0.49481548 0.4156446 0.3889514
#> [15,] 0.3213871 0.4090963 0.5129446 0.5125990 0.36720176 0.5454570 0.4967388
#> [16,] 0.5271038 0.5952604 0.7292842 0.8514803 0.80570861 0.8071103 0.7666732
#> [17,] 0.3980757 0.3078135 0.4351772 0.7655466 0.86699824 0.8127382 0.7772326
#> [18,] 0.3915051 0.2240605 0.3797742 0.9084254 0.94046444 0.5135525 0.3534723
#> [19,] 0.4390872 0.4708271 0.5167948 0.6740810 0.89423406 0.9514346 0.9233936
#> [20,] 0.2162316 0.1285429 0.2268909 0.2424036 0.14405687 0.2260401 0.3535004
#> [21,] 0.6540196 0.6547942 0.3755742 0.3429084 0.58165195 0.4976865 0.5173309
#> [22,] 0.4925000 0.8793000 0.9606000 0.8786000 0.69050000 0.6937000 0.5674000
#> [23,] 0.4743489 0.5851817 0.7297438 0.7561895 0.77805281 0.8711271 0.8689317
#> [24,] 0.2301455 0.1766232 0.1733341 0.2375113 0.07252064 0.2002558 0.4233404
#> [25,] 0.9273548 0.8859646 0.8458320 0.8733511 0.87363908 0.9093166 0.8556931
#>             V28       V29          V3        V30         V31        V32
#>  [1,] 0.9674100 0.8699262 0.034069630 0.66548738  0.47303906 0.33182315
#>  [2,] 0.7278000 0.7591000 0.057700000 0.65790000  0.75140000 0.66660000
#>  [3,] 0.1133679 0.1566608 0.123678566 0.14629049  0.23499495 0.28195166
#>  [4,] 0.3646385 0.3276804 0.008455016 0.55218246  0.68847395 0.79324904
#>  [5,] 0.9323893 0.9010275 0.023336012 0.93370179  0.91431381 0.82731528
#>  [6,] 0.8506405 0.7137319 0.030616176 0.55361939  0.40681175 0.32825429
#>  [7,] 0.6204119 0.5670433 0.021850461 0.41769542  0.31379286 0.19756371
#>  [8,] 0.4796530 0.6461656 0.062993335 0.54741565  0.20252661 0.08839523
#>  [9,] 0.8171919 0.8115990 0.081508527 0.84646980  0.79253729 0.68337099
#> [10,] 0.7286709 0.7048080 0.056269769 0.78231446  0.51345529 0.52833617
#> [11,] 0.4540943 0.4672275 0.071567852 0.40942417  0.40444746 0.16351259
#> [12,] 0.8488486 0.7590715 0.058565977 0.58697093  0.51437588 0.47462494
#> [13,] 0.2547577 0.1674911 0.019525739 0.08013329 -0.01430997 0.05961912
#> [14,] 0.4419819 0.4478643 0.059319524 0.41147844  0.24151633 0.26354505
#> [15,] 0.2798257 0.3251717 0.056225821 0.48515366  0.67793166 0.65567489
#> [16,] 0.7115121 0.6745109 0.062460027 0.65341744  0.51373643 0.27300158
#> [17,] 0.7583951 0.5082358 0.029412941 0.34676118  0.40605620 0.43792526
#> [18,] 0.6933035 0.8195373 0.061034984 0.45872097  0.53765587 0.69243407
#> [19,] 0.8171775 0.6954966 0.009615202 0.56765625  0.47726823 0.30133017
#> [20,] 0.5630990 0.6312992 0.031295704 0.66377585  0.67527166 0.58196113
#> [21,] 0.9300989 0.8461832 0.015226221 0.40901280  0.27182033 0.26641852
#> [22,] 0.6540000 0.7802000 0.039800000 0.75750000  0.58360000 0.63160000
#> [23,] 0.7973222 0.7539549 0.015291084 0.68660526  0.68320486 0.62694325
#> [24,] 0.6748937 0.8256286 0.018577418 0.74457143  0.51261439 0.19559444
#> [25,] 0.6716355 0.4489183 0.030162986 0.37759541  0.37833364 0.32336680
#>              V33        V34       V35        V36        V37       V38
#>  [1,] 0.29127154 0.24373165 0.1815303 0.09676960 0.11005920 0.1490040
#>  [2,] 0.49030000 0.59620000 0.6552000 0.40140000 0.11880000 0.3245000
#>  [3,] 0.32813184 0.46341927 0.5173907 0.29076107 0.21952125 0.2901958
#>  [4,] 0.95826988 0.91609320 0.6531982 0.37660384 0.08851826 0.1282542
#>  [5,] 0.71393585 0.64905469 0.6575612 0.70626910 0.68574634 0.5655016
#>  [6,] 0.25962809 0.18758109 0.1820515 0.23264915 0.23239658 0.2546582
#>  [7,] 0.04898717 0.09993943 0.2210051 0.26585633 0.18889416 0.1542205
#>  [8,] 0.22625215 0.34549767 0.3888888 0.28667410 0.31883028 0.4641662
#>  [9,] 0.42034082 0.37724876 0.3919717 0.28568832 0.31182775 0.3958055
#> [10,] 0.68991335 0.67271917 0.6680713 0.58222066 0.58736431 0.6204869
#> [11,] 0.48026720 0.67631236 0.6167783 0.44974535 0.42075042 0.3576400
#> [12,] 0.40406329 0.31985297 0.1834160 0.19532762 0.23477560 0.2230239
#> [13,] 0.19260838 0.08008533 0.1840382 0.05644977 0.23608058 0.2267694
#> [14,] 0.34647454 0.22821833 0.1919004 0.16754736 0.23792089 0.4531965
#> [15,] 0.54318050 0.38709733 0.4642015 0.40593032 0.55487548 0.8370831
#> [16,] 0.27124288 0.15473333 0.3198288 0.54921915 0.47457741 0.1837940
#> [17,] 0.33453307 0.21142682 0.3638417 0.39550387 0.19868312 0.1740945
#> [18,] 0.59197704 0.88602753 0.8039619 0.73971698 0.77487826 0.4733354
#> [19,] 0.22278582 0.17346410 0.1499660 0.21990492 0.18076439 0.2187612
#> [20,] 0.68679064 0.85248862 0.9130681 0.91410630 0.74983424 0.6317807
#> [21,] 0.10286395 0.15138792 0.3066027 0.20665883 0.22899964 0.2381047
#> [22,] 0.81080000 0.90390000 0.8647000 0.66950000 0.40270000 0.2370000
#> [23,] 0.55477174 0.57379510 0.5956124 0.54227177 0.42278751 0.3100139
#> [24,] 0.48278403 0.50056044 0.2559806 0.36404233 0.50574035 0.5859027
#> [25,] 0.29407168 0.29524240 0.2274750 0.12502218 0.19949982 0.1755844
#>              V39         V4        V40        V41        V42        V43
#>  [1,] 0.16395635 0.03046199 0.15134712 0.08666948 0.12162252 0.20631197
#>  [2,] 0.31070000 0.06270000 0.13540000 0.51090000 0.79880000 0.75170000
#>  [3,] 0.26941425 0.14131064 0.21419547 0.21485896 0.12452257 0.10891453
#>  [4,] 0.34302611 0.01960574 0.56691649 0.58635530 0.45449249 0.23636626
#>  [5,] 0.47941194 0.04954940 0.50127592 0.47778659 0.42478920 0.44703905
#>  [6,] 0.27387974 0.05551750 0.27778942 0.26974498 0.28982743 0.25623948
#>  [7,] 0.23091690 0.02769850 0.22321804 0.21273051 0.22462101 0.25316146
#>  [8,] 0.40552000 0.07153761 0.37028260 0.37632314 0.38996792 0.46315935
#>  [9,] 0.40928874 0.06515406 0.30901812 0.54935282 0.59266912 0.47013607
#> [10,] 0.54794677 0.09024535 0.37063449 0.34269983 0.36988843 0.36760079
#> [11,] 0.43014159 0.07374961 0.41301873 0.27767740 0.38048012 0.37616624
#> [12,] 0.27918248 0.07446522 0.33368108 0.31533312 0.34503671 0.26801038
#> [13,] 0.17300928 0.11559376 0.05235293 0.05746786 0.30201771 0.28667807
#> [14,] 0.37802571 0.12139519 0.21470189 0.10785483 0.20392684 0.27693477
#> [15,] 0.96410154 0.02079732 0.91747442 0.65900801 0.56342527 0.43945713
#> [16,] 0.15846071 0.06379010 0.21434015 0.34971754 0.32089179 0.23283264
#> [17,] 0.20786160 0.03641078 0.19663105 0.10206345 0.09285891 0.11544381
#> [18,] 0.33174887 0.08518119 0.32920688 0.27999808 0.27995320 0.05999947
#> [19,] 0.24325789 0.01782774 0.20686946 0.16591723 0.25751590 0.21807464
#> [20,] 0.58523587 0.03116578 0.61335826 0.57805264 0.51068693 0.34127162
#> [21,] 0.09910692 0.04782647 0.05174159 0.17368329 0.16495078 0.16923272
#> [22,] 0.26850000 0.03590000 0.36620000 0.32670000 0.22000000 0.29960000
#> [23,] 0.26859012 0.02231013 0.30347554 0.23116654 0.13583412 0.11394883
#> [24,] 0.62546264 0.02959816 0.46791680 0.22263370 0.17196533 0.11990983
#> [25,] 0.14345962 0.03692833 0.16811405 0.12070933 0.09417287 0.09717546
#>                V44         V45          V46         V47        V48         V49
#>  [1,]  0.206691952 0.178157264  0.181737102  0.15156927 0.09968268 0.055151587
#>  [2,]  0.550800000 0.585800000  0.729200000  0.55220000 0.33390000 0.160800000
#>  [3,]  0.084328628 0.034515646  0.048379555  0.10318837 0.09880872 0.069281334
#>  [4,]  0.091187417 0.122020161  0.139108461  0.14026732 0.13676164 0.087347243
#>  [5,]  0.516556878 0.476290280  0.335760739  0.23645034 0.19153938 0.104741043
#>  [6,]  0.252340925 0.204306233  0.120545926  0.10089256 0.07738177 0.045538433
#>  [7,]  0.236338494 0.116545163  0.075218818  0.09497383 0.11636862 0.081420552
#>  [8,]  0.354760953 0.170056312  0.111859838  0.13517775 0.05225183 0.033621776
#>  [9,]  0.406641324 0.531241134  0.502801036  0.33863615 0.20867244 0.130465784
#> [10,]  0.279396784 0.309948731  0.252188993  0.07877894 0.06112288 0.037906970
#> [11,]  0.191220266 0.150329420  0.033463078  0.11142510 0.18532304 0.093702607
#> [12,]  0.193773117 0.218481786  0.186654969  0.17197952 0.14743934 0.077011446
#> [13,] -0.006382838 0.221422397  0.164842746  0.04148472 0.06844900 0.058323526
#> [14,]  0.144341174 0.059214899 -0.011048039 -0.01342889 0.03393666 0.011722469
#> [15,]  0.439295370 0.473846277  0.388442477  0.27879178 0.16523130 0.074656597
#> [16,]  0.145912700 0.152187858  0.154577100  0.10846410 0.09230144 0.045140358
#> [17,]  0.125535613 0.113822789  0.087961643  0.04223495 0.04117072 0.012487774
#> [18,]  0.176240884 0.151225386  0.162938646  0.12269064 0.07556888 0.042502245
#> [19,]  0.145211869 0.109355178  0.083669657  0.07838001 0.06720765 0.032565012
#> [20,]  0.216330717 0.184040378  0.160758633  0.14761112 0.11341085 0.069007823
#> [21,]  0.168968434 0.097301278  0.062402207  0.01596220 0.03081586 0.008945374
#> [22,]  0.220500000 0.116300000  0.063500000  0.04650000 0.04220000 0.017400000
#> [23,]  0.092843781 0.052138969  0.051414455  0.03559817 0.03536053 0.010587475
#> [24,]  0.044087384 0.002398209 -0.008388603  0.05027644 0.05562841 0.013897110
#> [25,]  0.125859403 0.119588265  0.090629742  0.08744189 0.05938331 0.036694311
#>                V5         V50         V51          V52          V53
#>  [1,] 0.055369752 0.015889996 0.016372013  0.011918578 0.0057147569
#>  [2,] 0.063500000 0.047500000 0.100400000  0.070900000 0.0317000000
#>  [3,] 0.135500311 0.027009973 0.018839532  0.036855805 0.0146367481
#>  [4,] 0.038742207 0.022245734 0.010570026  0.004537392 0.0008761102
#>  [5,] 0.077934557 0.029656470 0.017012968  0.013846875 0.0100216159
#>  [6,] 0.098399209 0.017775502 0.017470357  0.015325617 0.0140529956
#>  [7,] 0.088007582 0.018654185 0.027721017  0.022600890 0.0167041641
#>  [8,] 0.062503783 0.015549743 0.001649088 -0.011483384 0.0107533345
#>  [9,] 0.078545999 0.049408465 0.041573783  0.025939726 0.0121975084
#> [10,] 0.144764910 0.015835539 0.013275391  0.022456324 0.0140055387
#> [11,] 0.016976769 0.020579461 0.018234558  0.016211036 0.0164949108
#> [12,] 0.080940331 0.016940635 0.021427172  0.016613888 0.0102459113
#> [13,] 0.267886026 0.059306084 0.028960019  0.034241278 0.0263388811
#> [14,] 0.159515220 0.007279953 0.006370958  0.013611360 0.0097849877
#> [15,] 0.048556790 0.026833484 0.023247016  0.011523842 0.0065550725
#> [16,] 0.063211860 0.019438843 0.014011508  0.013313902 0.0115922359
#> [17,] 0.090570465 0.019467813 0.005365367  0.023025634 0.0099858003
#> [18,] 0.148834272 0.007944983 0.017078606  0.004640388 0.0146170832
#> [19,] 0.024894776 0.010100655 0.005514281  0.005169083 0.0038635054
#> [20,] 0.056269260 0.026806402 0.017405966  0.008747170 0.0122018262
#> [21,] 0.007535407 0.010550311 0.009974011  0.002439051 0.0079962072
#> [22,] 0.068100000 0.017200000 0.013400000  0.014100000 0.0191000000
#> [23,] 0.055008749 0.008164922 0.006601583  0.003836186 0.0053147314
#> [24,] 0.044107079 0.008007951 0.004352006  0.017791691 0.0269747390
#> [25,] 0.067836588 0.018222833 0.012505190  0.010551128 0.0090079162
#>                V54         V55         V56         V57         V58         V59
#>  [1,] 0.0060248900 0.004754511 0.003971855 0.005100093 0.004071058 0.004284413
#>  [2,] 0.0309000000 0.025200000 0.008700000 0.017700000 0.021400000 0.022700000
#>  [3,] 0.0108675600 0.021153676 0.013499531 0.003921259 0.011699778 0.022404958
#>  [4,] 0.0006782778 0.007294045 0.004575253 0.002846686 0.006282319 0.006876628
#>  [5,] 0.0070719492 0.005530683 0.007286168 0.006072913 0.004726723 0.002807652
#>  [6,] 0.0133225442 0.006326763 0.009813695 0.007571182 0.005768718 0.007347927
#>  [7,] 0.0214476954 0.015970799 0.005950426 0.005322878 0.019787737 0.011430358
#>  [8,] 0.0272231173 0.007135635 0.019696563 0.009416386 0.004170539 0.018412145
#>  [9,] 0.0166831961 0.015309743 0.013045981 0.009842321 0.012330291 0.013093814
#> [10,] 0.0129035441 0.008505006 0.005573853 0.001524712 0.011653189 0.012519005
#> [11,] 0.0211330004 0.013910224 0.006397513 0.003690379 0.016320462 0.008456765
#> [12,] 0.0106561625 0.009263545 0.007260599 0.006804958 0.008244301 0.007335297
#> [13,] 0.0168934637 0.023107843 0.005039102 0.003218529 0.013771787 0.017894227
#> [14,] 0.0149506281 0.022315953 0.029420601 0.006928332 0.005224500 0.016581902
#> [15,] 0.0133859224 0.009536115 0.010326215 0.004245303 0.008174623 0.009304553
#> [16,] 0.0107500507 0.010360339 0.009012409 0.007388816 0.005691671 0.008092650
#> [17,] 0.0147023638 0.011872274 0.005552495 0.005657928 0.003311805 0.008184825
#> [18,] 0.0067340435 0.019549443 0.007999242 0.014543681 0.005488542 0.018639603
#> [19,] 0.0046782476 0.004397544 0.004060337 0.003661638 0.003387064 0.002797845
#> [20,] 0.0126515638 0.010279198 0.008250147 0.009765072 0.009993977 0.011098406
#> [21,] 0.0142404222 0.018775087 0.012018416 0.009390759 0.008075505 0.006045897
#> [22,] 0.0145000000 0.006500000 0.012900000 0.021700000 0.008700000 0.007700000
#> [23,] 0.0078793884 0.008023495 0.008589875 0.007654692 0.005185697 0.005874651
#> [24,] 0.0168889293 0.019382998 0.005859758 0.005129536 0.005457779 0.010154656
#> [25,] 0.0063379302 0.005788977 0.008549682 0.006868111 0.004342426 0.005196099
#>               V6         V60          V7         V8          V9
#>  [1,] 0.08030854 0.004793895 0.115662813 0.14226591  0.20020734
#>  [2,] 0.13280000 0.010600000 0.098800000 0.17870000  0.11990000
#>  [3,] 0.05049410 0.011356064 0.073352416 0.05060247  0.12930820
#>  [4,] 0.07352335 0.002735225 0.069692385 0.14908961  0.22485831
#>  [5,] 0.10746195 0.003131348 0.137609728 0.14777078  0.27851825
#>  [6,] 0.11812660 0.005264367 0.140124202 0.15770034  0.19494246
#>  [7,] 0.19152096 0.003960338 0.187172642 0.14171574  0.15322013
#>  [8,] 0.03731606 0.012977963 0.009753383 0.06837056 -0.01618771
#>  [9,] 0.10250100 0.010257357 0.109978010 0.15630645  0.21452371
#> [10,] 0.10463025 0.008622861 0.104892200 0.26145157  0.48516595
#> [11,] 0.02758435 0.003311516 0.085644951 0.24049340  0.42805782
#> [12,] 0.09506081 0.005131518 0.112092474 0.11752884  0.15686871
#> [13,] 0.30050364 0.011184857 0.220337349 0.14500509  0.14955797
#> [14,] 0.07065841 0.010747074 0.087973908 0.10750738  0.03507963
#> [15,] 0.06024131 0.013435508 0.123185247 0.14578642  0.05881753
#> [16,] 0.09737569 0.004856872 0.128829802 0.14237393  0.15401049
#> [17,] 0.10066704 0.004922893 0.067273733 0.05513082  0.10616624
#> [18,] 0.21831435 0.007468485 0.206694992 0.18281356  0.13237502
#> [19,] 0.04548343 0.003576675 0.069609538 0.06275430  0.07741883
#> [20,] 0.09889590 0.006420055 0.110313064 0.08236008  0.02691677
#> [21,] 0.07173815 0.008017702 0.140183105 0.16651749  0.22009420
#> [22,] 0.07060000 0.012200000 0.102000000 0.08930000  0.03810000
#> [23,] 0.10037061 0.004044575 0.101340889 0.09771812  0.11331568
#> [24,] 0.06334843 0.009649500 0.026850154 0.06285723  0.13004479
#> [25,] 0.10128524 0.004106158 0.130192306 0.10758344  0.13775996
#> 
#> $cl
#>  [1] M M M M M M M M M M M M M M R R R R R R R R R R R
#> Levels: M R
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913