Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0348  0.0213
#>   std. dev.      0.0258   0.016
#>   weight sum         73      66
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2527  0.1608
#>   std. dev.      0.1382  0.1107
#>   weight sum         73      66
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean           0.2898  0.1671
#>   std. dev.      0.1306  0.1085
#>   weight sum         73      66
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.2997  0.1844
#>   std. dev.      0.1231  0.1275
#>   weight sum         73      66
#>   precision      0.0046  0.0046
#> 
#> V13
#>   mean           0.3132  0.2192
#>   std. dev.      0.1269   0.131
#>   weight sum         73      66
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3124   0.257
#>   std. dev.      0.1443  0.1493
#>   weight sum         73      66
#>   precision      0.0056  0.0056
#> 
#> V15
#>   mean           0.3222  0.2984
#>   std. dev.      0.1841  0.1947
#>   weight sum         73      66
#>   precision      0.0057  0.0057
#> 
#> V16
#>   mean           0.3823   0.372
#>   std. dev.      0.2072  0.2347
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V17
#>   mean           0.4208   0.408
#>   std. dev.      0.2368   0.274
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4606  0.4288
#>   std. dev.      0.2546  0.2546
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5397  0.4528
#>   std. dev.      0.2579  0.2485
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V2
#>   mean           0.0444  0.0284
#>   std. dev.       0.038    0.02
#>   weight sum         73      66
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6061  0.4991
#>   std. dev.      0.2682  0.2673
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean            0.657    0.54
#>   std. dev.      0.2622  0.2571
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6759  0.5468
#>   std. dev.      0.2511  0.2777
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6835  0.5879
#>   std. dev.      0.2587  0.2563
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.7002  0.6312
#>   std. dev.      0.2421  0.2337
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V25
#>   mean           0.6863  0.6498
#>   std. dev.      0.2342  0.2601
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.6904  0.6873
#>   std. dev.      0.2385  0.2486
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V27
#>   mean             0.69  0.7025
#>   std. dev.      0.2722  0.2182
#>   weight sum         73      66
#>   precision      0.0076  0.0076
#> 
#> V28
#>   mean           0.6855  0.7049
#>   std. dev.      0.2619  0.1851
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V29
#>   mean           0.6255  0.6481
#>   std. dev.      0.2487  0.2265
#>   weight sum         73      66
#>   precision      0.0075  0.0075
#> 
#> V3
#>   mean           0.0498  0.0345
#>   std. dev.      0.0456  0.0262
#>   weight sum         73      66
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5738  0.5708
#>   std. dev.      0.2163  0.2398
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V31
#>   mean            0.489  0.5172
#>   std. dev.      0.2225  0.2067
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V32
#>   mean           0.4446  0.4403
#>   std. dev.      0.2129  0.2122
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4304  0.4277
#>   std. dev.      0.1968  0.2121
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3939  0.4482
#>   std. dev.      0.2166  0.2498
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3577  0.4554
#>   std. dev.      0.2413  0.2722
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3301  0.4601
#>   std. dev.      0.2435  0.2677
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3204  0.4095
#>   std. dev.      0.2266  0.2464
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean           0.3288  0.3412
#>   std. dev.      0.1984  0.2239
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V39
#>   mean            0.322   0.324
#>   std. dev.      0.1734  0.2099
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V4
#>   mean           0.0683  0.0415
#>   std. dev.      0.0563  0.0315
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3002  0.3106
#>   std. dev.      0.1674  0.1958
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean            0.293   0.265
#>   std. dev.      0.1725  0.1693
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.2886  0.2297
#>   std. dev.      0.1593  0.1522
#>   weight sum         73      66
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2658  0.2029
#>   std. dev.      0.1232  0.1127
#>   weight sum         73      66
#>   precision      0.0043  0.0043
#> 
#> V44
#>   mean           0.2417  0.1654
#>   std. dev.      0.1365  0.0878
#>   weight sum         73      66
#>   precision      0.0043  0.0043
#> 
#> V45
#>   mean           0.2298  0.1287
#>   std. dev.      0.1626  0.0768
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V46
#>   mean           0.1842   0.099
#>   std. dev.      0.1427   0.083
#>   weight sum         73      66
#>   precision      0.0046  0.0046
#> 
#> V47
#>   mean            0.137  0.0821
#>   std. dev.      0.0878  0.0608
#>   weight sum         73      66
#>   precision      0.0032  0.0032
#> 
#> V48
#>   mean           0.1049   0.063
#>   std. dev.       0.064  0.0473
#>   weight sum         73      66
#>   precision      0.0022  0.0022
#> 
#> V49
#>   mean           0.0601  0.0352
#>   std. dev.      0.0352    0.03
#>   weight sum         73      66
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0844  0.0638
#>   std. dev.      0.0604  0.0488
#>   weight sum         73      66
#>   precision      0.0031  0.0031
#> 
#> V50
#>   mean           0.0218  0.0172
#>   std. dev.      0.0142  0.0117
#>   weight sum         73      66
#>   precision      0.0008  0.0008
#> 
#> V51
#>   mean           0.0182  0.0119
#>   std. dev.       0.012  0.0081
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V52
#>   mean           0.0152  0.0106
#>   std. dev.      0.0099  0.0066
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0113  0.0101
#>   std. dev.      0.0078  0.0063
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean            0.013  0.0101
#>   std. dev.      0.0089  0.0056
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0095  0.0087
#>   std. dev.      0.0082  0.0048
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V56
#>   mean           0.0097  0.0072
#>   std. dev.      0.0072  0.0045
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0079  0.0078
#>   std. dev.      0.0065  0.0054
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0095  0.0063
#>   std. dev.       0.008  0.0045
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0094  0.0072
#>   std. dev.      0.0075  0.0049
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean            0.108  0.1013
#>   std. dev.      0.0511  0.0719
#>   weight sum         73      66
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean           0.0073   0.006
#>   std. dev.      0.0065  0.0034
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean            0.128  0.1138
#>   std. dev.      0.0571  0.0707
#>   weight sum         73      66
#>   precision      0.0029  0.0029
#> 
#> V8
#>   mean           0.1473  0.1167
#>   std. dev.      0.0882  0.0824
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2183  0.1351
#>   std. dev.      0.1225  0.1008
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3478261