Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                  (0.5)   (0.5)
#> ===============================
#> V1
#>   mean           0.0346  0.0233
#>   std. dev.      0.0218  0.0157
#>   weight sum         69      70
#>   precision      0.0009  0.0009
#> 
#> V10
#>   mean           0.2513  0.1551
#>   std. dev.      0.1411  0.1216
#>   weight sum         69      70
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean           0.2922  0.1767
#>   std. dev.      0.1356  0.1226
#>   weight sum         69      70
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.3063  0.2011
#>   std. dev.       0.135  0.1415
#>   weight sum         69      70
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3211  0.2306
#>   std. dev.      0.1287  0.1388
#>   weight sum         69      70
#>   precision      0.0049  0.0049
#> 
#> V14
#>   mean           0.3255  0.2642
#>   std. dev.      0.1663  0.1659
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3307  0.3003
#>   std. dev.      0.1996  0.2118
#>   weight sum         69      70
#>   precision      0.0074  0.0074
#> 
#> V16
#>   mean           0.3858  0.3765
#>   std. dev.      0.2165  0.2515
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V17
#>   mean           0.4275  0.4227
#>   std. dev.       0.239  0.2861
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4693  0.4581
#>   std. dev.      0.2587  0.2743
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5535  0.4874
#>   std. dev.      0.2598  0.2608
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0423  0.0302
#>   std. dev.      0.0293  0.0188
#>   weight sum         69      70
#>   precision       0.001   0.001
#> 
#> V20
#>   mean           0.6222  0.5097
#>   std. dev.      0.2693  0.2686
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6666  0.5342
#>   std. dev.      0.2553  0.2503
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6863  0.5447
#>   std. dev.      0.2352  0.2524
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6998    0.59
#>   std. dev.      0.2433  0.2294
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6939  0.6416
#>   std. dev.      0.2468  0.2213
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V25
#>   mean           0.6716  0.6552
#>   std. dev.      0.2506  0.2492
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.6844  0.6697
#>   std. dev.      0.2548  0.2438
#>   weight sum         69      70
#>   precision      0.0068  0.0068
#> 
#> V27
#>   mean           0.6888    0.67
#>   std. dev.      0.2836  0.2184
#>   weight sum         69      70
#>   precision      0.0074  0.0074
#> 
#> V28
#>   mean           0.6908  0.6653
#>   std. dev.      0.2725  0.2107
#>   weight sum         69      70
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean            0.632  0.6326
#>   std. dev.      0.2562  0.2383
#>   weight sum         69      70
#>   precision      0.0075  0.0075
#> 
#> V3
#>   mean           0.0498  0.0345
#>   std. dev.      0.0385  0.0246
#>   weight sum         69      70
#>   precision      0.0016  0.0016
#> 
#> V30
#>   mean           0.5657  0.5844
#>   std. dev.      0.2057  0.2452
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V31
#>   mean           0.4759  0.5398
#>   std. dev.      0.2117   0.215
#>   weight sum         69      70
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4294  0.4679
#>   std. dev.      0.2121  0.2191
#>   weight sum         69      70
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4056  0.4554
#>   std. dev.      0.1936   0.214
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3629  0.4674
#>   std. dev.      0.2106  0.2614
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean             0.32  0.4822
#>   std. dev.      0.2391  0.2819
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.2925  0.4878
#>   std. dev.       0.226  0.2722
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V37
#>   mean           0.2799  0.4324
#>   std. dev.      0.2126  0.2496
#>   weight sum         69      70
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3109  0.3577
#>   std. dev.      0.1868  0.2256
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3187   0.323
#>   std. dev.      0.1809  0.1994
#>   weight sum         69      70
#>   precision      0.0062  0.0062
#> 
#> V4
#>   mean           0.0657  0.0403
#>   std. dev.      0.0445  0.0305
#>   weight sum         69      70
#>   precision       0.002   0.002
#> 
#> V40
#>   mean           0.2897  0.3154
#>   std. dev.      0.1707  0.1769
#>   weight sum         69      70
#>   precision      0.0064  0.0064
#> 
#> V41
#>   mean           0.2837  0.2817
#>   std. dev.      0.1771  0.1563
#>   weight sum         69      70
#>   precision      0.0054  0.0054
#> 
#> V42
#>   mean           0.2945  0.2505
#>   std. dev.      0.1724   0.148
#>   weight sum         69      70
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean            0.271  0.2149
#>   std. dev.      0.1356  0.1222
#>   weight sum         69      70
#>   precision      0.0057  0.0057
#> 
#> V44
#>   mean           0.2497  0.1854
#>   std. dev.      0.1403  0.1113
#>   weight sum         69      70
#>   precision      0.0058  0.0058
#> 
#> V45
#>   mean           0.2375  0.1449
#>   std. dev.      0.1654  0.0877
#>   weight sum         69      70
#>   precision      0.0048  0.0048
#> 
#> V46
#>   mean           0.1903  0.1134
#>   std. dev.      0.1514  0.0772
#>   weight sum         69      70
#>   precision      0.0054  0.0054
#> 
#> V47
#>   mean           0.1442  0.0896
#>   std. dev.      0.0988  0.0514
#>   weight sum         69      70
#>   precision      0.0041  0.0041
#> 
#> V48
#>   mean           0.1121  0.0659
#>   std. dev.      0.0695    0.04
#>   weight sum         69      70
#>   precision      0.0024  0.0024
#> 
#> V49
#>   mean           0.0629  0.0377
#>   std. dev.       0.035  0.0236
#>   weight sum         69      70
#>   precision      0.0012  0.0012
#> 
#> V5
#>   mean           0.0916  0.0625
#>   std. dev.      0.0564  0.0467
#>   weight sum         69      70
#>   precision      0.0024  0.0024
#> 
#> V50
#>   mean           0.0218  0.0175
#>   std. dev.      0.0129  0.0113
#>   weight sum         69      70
#>   precision      0.0008  0.0008
#> 
#> V51
#>   mean           0.0193  0.0118
#>   std. dev.      0.0142  0.0075
#>   weight sum         69      70
#>   precision      0.0009  0.0009
#> 
#> V52
#>   mean           0.0171  0.0097
#>   std. dev.      0.0115  0.0062
#>   weight sum         69      70
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0119  0.0097
#>   std. dev.      0.0086  0.0058
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0122  0.0097
#>   std. dev.      0.0088  0.0052
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean            0.011  0.0087
#>   std. dev.      0.0094  0.0047
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0091  0.0072
#>   std. dev.      0.0071  0.0049
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0084  0.0085
#>   std. dev.      0.0067  0.0063
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0092   0.007
#>   std. dev.      0.0075   0.005
#>   weight sum         69      70
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0092  0.0073
#>   std. dev.      0.0074  0.0054
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1149  0.0959
#>   std. dev.      0.0559   0.063
#>   weight sum         69      70
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0068   0.006
#>   std. dev.      0.0061  0.0038
#>   weight sum         69      70
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1314  0.1126
#>   std. dev.      0.0593  0.0621
#>   weight sum         69      70
#>   precision      0.0025  0.0025
#> 
#> V8
#>   mean           0.1485  0.1119
#>   std. dev.      0.0825  0.0821
#>   weight sum         69      70
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2074  0.1325
#>   std. dev.       0.114  0.1084
#>   weight sum         69      70
#>   precision      0.0047  0.0047
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3913043