Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.53)  (0.47)
#> ===============================
#> V1
#>   mean           0.0349  0.0224
#>   std. dev.      0.0254  0.0151
#>   weight sum         74      65
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2545  0.1707
#>   std. dev.      0.1419  0.1203
#>   weight sum         74      65
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean            0.296  0.1784
#>   std. dev.       0.124  0.1163
#>   weight sum         74      65
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.3067  0.1926
#>   std. dev.      0.1206  0.1327
#>   weight sum         74      65
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3206  0.2267
#>   std. dev.      0.1267  0.1407
#>   weight sum         74      65
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3413  0.2672
#>   std. dev.      0.1612  0.1681
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3547  0.3054
#>   std. dev.       0.194   0.218
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V16
#>   mean           0.4075  0.3822
#>   std. dev.      0.2064  0.2468
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V17
#>   mean            0.438  0.4279
#>   std. dev.      0.2436  0.2839
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4737  0.4529
#>   std. dev.      0.2625  0.2602
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean           0.5489  0.4627
#>   std. dev.      0.2598  0.2503
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V2
#>   mean           0.0472  0.0317
#>   std. dev.      0.0398  0.0265
#>   weight sum         74      65
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6279  0.4792
#>   std. dev.      0.2533  0.2432
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V21
#>   mean           0.6761  0.5259
#>   std. dev.      0.2497  0.2383
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V22
#>   mean           0.6715  0.5612
#>   std. dev.      0.2428  0.2541
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V23
#>   mean           0.6731  0.5919
#>   std. dev.      0.2629  0.2361
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V24
#>   mean           0.6824  0.6356
#>   std. dev.      0.2527  0.2212
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6644  0.6675
#>   std. dev.      0.2571    0.22
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V26
#>   mean           0.6848  0.6842
#>   std. dev.      0.2539   0.224
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V27
#>   mean           0.6935  0.6628
#>   std. dev.      0.2756  0.2238
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V28
#>   mean           0.6896  0.6612
#>   std. dev.       0.271  0.2157
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6289  0.6361
#>   std. dev.      0.2518   0.248
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V3
#>   mean           0.0519  0.0365
#>   std. dev.      0.0448  0.0293
#>   weight sum         74      65
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5753  0.5882
#>   std. dev.      0.2157  0.2402
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4749  0.5392
#>   std. dev.      0.2255  0.2121
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean            0.423  0.4762
#>   std. dev.      0.2187  0.2177
#>   weight sum         74      65
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.3903  0.4663
#>   std. dev.      0.2063  0.2267
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V34
#>   mean           0.3721  0.4581
#>   std. dev.      0.2203   0.253
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean           0.3506   0.473
#>   std. dev.      0.2442  0.2525
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3181  0.4826
#>   std. dev.      0.2418  0.2595
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V37
#>   mean           0.3099  0.4446
#>   std. dev.      0.2164  0.2491
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean            0.329  0.3666
#>   std. dev.      0.1995  0.2283
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V39
#>   mean           0.3362   0.338
#>   std. dev.      0.1784  0.2145
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0686  0.0433
#>   std. dev.      0.0564  0.0351
#>   weight sum         74      65
#>   precision      0.0034  0.0034
#> 
#> V40
#>   mean           0.3082  0.3354
#>   std. dev.      0.1635  0.2007
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2924  0.3031
#>   std. dev.      0.1735  0.1716
#>   weight sum         74      65
#>   precision      0.0062  0.0062
#> 
#> V42
#>   mean           0.2868  0.2623
#>   std. dev.      0.1657  0.1583
#>   weight sum         74      65
#>   precision      0.0058  0.0058
#> 
#> V43
#>   mean           0.2612  0.2159
#>   std. dev.      0.1301  0.1287
#>   weight sum         74      65
#>   precision      0.0056  0.0056
#> 
#> V44
#>   mean            0.229  0.1871
#>   std. dev.       0.133  0.1182
#>   weight sum         74      65
#>   precision      0.0059  0.0059
#> 
#> V45
#>   mean            0.229  0.1583
#>   std. dev.      0.1646  0.1044
#>   weight sum         74      65
#>   precision      0.0047  0.0047
#> 
#> V46
#>   mean           0.1877  0.1311
#>   std. dev.      0.1448  0.0998
#>   weight sum         74      65
#>   precision      0.0046  0.0046
#> 
#> V47
#>   mean           0.1379  0.0993
#>   std. dev.      0.0778  0.0747
#>   weight sum         74      65
#>   precision      0.0032  0.0032
#> 
#> V48
#>   mean           0.1059  0.0704
#>   std. dev.      0.0565  0.0531
#>   weight sum         74      65
#>   precision      0.0022  0.0022
#> 
#> V49
#>   mean           0.0629  0.0398
#>   std. dev.      0.0331  0.0326
#>   weight sum         74      65
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0891  0.0649
#>   std. dev.      0.0612  0.0521
#>   weight sum         74      65
#>   precision      0.0029  0.0029
#> 
#> V50
#>   mean           0.0238  0.0178
#>   std. dev.      0.0153   0.013
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean            0.019  0.0128
#>   std. dev.      0.0118  0.0081
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V52
#>   mean           0.0158  0.0103
#>   std. dev.      0.0088  0.0063
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0123  0.0093
#>   std. dev.      0.0076  0.0062
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0131  0.0094
#>   std. dev.      0.0083  0.0051
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0112  0.0087
#>   std. dev.      0.0092  0.0051
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0098  0.0075
#>   std. dev.       0.007  0.0051
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0082  0.0083
#>   std. dev.      0.0063  0.0059
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0101   0.007
#>   std. dev.       0.008  0.0049
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0093  0.0075
#>   std. dev.       0.007  0.0056
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1125  0.0971
#>   std. dev.      0.0524  0.0687
#>   weight sum         74      65
#>   precision      0.0027  0.0027
#> 
#> V60
#>   mean           0.0075  0.0059
#>   std. dev.      0.0064  0.0037
#>   weight sum         74      65
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1334  0.1153
#>   std. dev.       0.058  0.0714
#>   weight sum         74      65
#>   precision      0.0028  0.0028
#> 
#> V8
#>   mean           0.1522  0.1211
#>   std. dev.      0.0914  0.0866
#>   weight sum         74      65
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2172  0.1417
#>   std. dev.      0.1203    0.11
#>   weight sum         74      65
#>   precision      0.0048  0.0048
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478