Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.53)  (0.47)
#> ===============================
#> V1
#>   mean           0.0347  0.0226
#>   std. dev.      0.0265   0.014
#>   weight sum         74      65
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2492  0.1663
#>   std. dev.      0.1306  0.1218
#>   weight sum         74      65
#>   precision      0.0047  0.0047
#> 
#> V11
#>   mean           0.2862  0.1854
#>   std. dev.      0.1102  0.1194
#>   weight sum         74      65
#>   precision      0.0045  0.0045
#> 
#> V12
#>   mean           0.2961  0.2069
#>   std. dev.      0.1202  0.1417
#>   weight sum         74      65
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3161  0.2355
#>   std. dev.      0.1329  0.1369
#>   weight sum         74      65
#>   precision      0.0054  0.0054
#> 
#> V14
#>   mean           0.3388  0.2616
#>   std. dev.      0.1734  0.1521
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3465  0.2915
#>   std. dev.      0.2137  0.1969
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V16
#>   mean            0.398  0.3694
#>   std. dev.      0.2325    0.25
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V17
#>   mean           0.4348  0.4185
#>   std. dev.      0.2621   0.284
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4792  0.4529
#>   std. dev.      0.2741  0.2654
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean           0.5535  0.4838
#>   std. dev.      0.2722   0.255
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V2
#>   mean           0.0456  0.0284
#>   std. dev.      0.0328  0.0192
#>   weight sum         74      65
#>   precision      0.0012  0.0012
#> 
#> V20
#>   mean           0.6398  0.5305
#>   std. dev.      0.2566  0.2532
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V21
#>   mean           0.6791  0.5568
#>   std. dev.      0.2567  0.2539
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6633  0.5857
#>   std. dev.      0.2443  0.2503
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V23
#>   mean            0.653   0.634
#>   std. dev.      0.2559  0.2319
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V24
#>   mean           0.6549  0.6638
#>   std. dev.       0.248  0.2395
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6528  0.6831
#>   std. dev.      0.2485  0.2492
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V26
#>   mean            0.693  0.7047
#>   std. dev.      0.2375  0.2432
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V27
#>   mean           0.7094  0.7021
#>   std. dev.      0.2678  0.2269
#>   weight sum         74      65
#>   precision      0.0076  0.0076
#> 
#> V28
#>   mean           0.7057   0.674
#>   std. dev.      0.2625  0.2005
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6473   0.619
#>   std. dev.      0.2539  0.2325
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V3
#>   mean           0.0499  0.0364
#>   std. dev.      0.0378  0.0243
#>   weight sum         74      65
#>   precision      0.0015  0.0015
#> 
#> V30
#>   mean           0.5913  0.5635
#>   std. dev.      0.2255  0.2395
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V31
#>   mean           0.4793  0.5193
#>   std. dev.      0.2416  0.1994
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4316  0.4649
#>   std. dev.      0.2267  0.2051
#>   weight sum         74      65
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4052  0.4674
#>   std. dev.      0.2022  0.2147
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V34
#>   mean           0.3835  0.4785
#>   std. dev.      0.2095  0.2513
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean           0.3639  0.4753
#>   std. dev.       0.251  0.2755
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3339  0.4643
#>   std. dev.      0.2661  0.2757
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3366  0.4269
#>   std. dev.        0.24  0.2492
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3507  0.3568
#>   std. dev.      0.2164  0.2428
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3488  0.3231
#>   std. dev.      0.1867  0.2165
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0606  0.0417
#>   std. dev.      0.0445  0.0297
#>   weight sum         74      65
#>   precision       0.002   0.002
#> 
#> V40
#>   mean            0.307  0.3185
#>   std. dev.      0.1568  0.1979
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V41
#>   mean           0.2939  0.2829
#>   std. dev.      0.1624  0.1835
#>   weight sum         74      65
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.3005  0.2535
#>   std. dev.      0.1668  0.1731
#>   weight sum         74      65
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean            0.276  0.2116
#>   std. dev.      0.1356  0.1266
#>   weight sum         74      65
#>   precision      0.0045  0.0045
#> 
#> V44
#>   mean           0.2438  0.1759
#>   std. dev.      0.1468   0.092
#>   weight sum         74      65
#>   precision      0.0042  0.0042
#> 
#> V45
#>   mean           0.2464  0.1411
#>   std. dev.      0.1765  0.0852
#>   weight sum         74      65
#>   precision      0.0047  0.0047
#> 
#> V46
#>   mean           0.2049  0.1176
#>   std. dev.      0.1495  0.0956
#>   weight sum         74      65
#>   precision      0.0045  0.0045
#> 
#> V47
#>   mean           0.1478  0.0936
#>   std. dev.      0.0885  0.0672
#>   weight sum         74      65
#>   precision      0.0031  0.0031
#> 
#> V48
#>   mean           0.1091  0.0657
#>   std. dev.      0.0632  0.0439
#>   weight sum         74      65
#>   precision      0.0021  0.0021
#> 
#> V49
#>   mean           0.0634  0.0353
#>   std. dev.      0.0348  0.0273
#>   weight sum         74      65
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0869  0.0646
#>   std. dev.      0.0553  0.0485
#>   weight sum         74      65
#>   precision      0.0024  0.0024
#> 
#> V50
#>   mean            0.024  0.0181
#>   std. dev.      0.0152  0.0124
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0192  0.0126
#>   std. dev.      0.0119  0.0094
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V52
#>   mean           0.0152   0.011
#>   std. dev.      0.0081  0.0078
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0113  0.0101
#>   std. dev.      0.0076  0.0059
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0125  0.0094
#>   std. dev.      0.0088  0.0051
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0106  0.0092
#>   std. dev.      0.0092   0.005
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0094  0.0074
#>   std. dev.      0.0071   0.005
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0077  0.0076
#>   std. dev.       0.006  0.0057
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0093  0.0064
#>   std. dev.      0.0082  0.0047
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean            0.008  0.0074
#>   std. dev.      0.0063   0.005
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1105  0.0989
#>   std. dev.      0.0512  0.0651
#>   weight sum         74      65
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0064  0.0061
#>   std. dev.      0.0044  0.0034
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V7
#>   mean           0.1254  0.1139
#>   std. dev.      0.0592   0.062
#>   weight sum         74      65
#>   precision      0.0025  0.0025
#> 
#> V8
#>   mean           0.1527  0.1141
#>   std. dev.      0.0918    0.08
#>   weight sum         74      65
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2193  0.1383
#>   std. dev.      0.1232  0.1087
#>   weight sum         74      65
#>   precision      0.0048  0.0048
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3623188