Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.55)  (0.45)
#> ===============================
#> V1
#>   mean           0.0366  0.0237
#>   std. dev.      0.0293  0.0152
#>   weight sum         77      62
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2569  0.1575
#>   std. dev.      0.1516  0.1155
#>   weight sum         77      62
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean           0.2903  0.1774
#>   std. dev.      0.1376  0.1217
#>   weight sum         77      62
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.2949  0.1963
#>   std. dev.      0.1283  0.1427
#>   weight sum         77      62
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3019  0.2297
#>   std. dev.      0.1245  0.1418
#>   weight sum         77      62
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3064  0.2704
#>   std. dev.      0.1395  0.1653
#>   weight sum         77      62
#>   precision      0.0055  0.0055
#> 
#> V15
#>   mean           0.3182  0.3007
#>   std. dev.       0.184  0.2076
#>   weight sum         77      62
#>   precision      0.0063  0.0063
#> 
#> V16
#>   mean           0.3807  0.3628
#>   std. dev.      0.2046  0.2423
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V17
#>   mean           0.4064  0.4023
#>   std. dev.      0.2351  0.2808
#>   weight sum         77      62
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4536  0.4383
#>   std. dev.      0.2449   0.268
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5455  0.4777
#>   std. dev.      0.2478  0.2583
#>   weight sum         77      62
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0469  0.0312
#>   std. dev.      0.0408  0.0256
#>   weight sum         77      62
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6248  0.5026
#>   std. dev.      0.2607  0.2612
#>   weight sum         77      62
#>   precision      0.0066  0.0066
#> 
#> V21
#>   mean            0.671   0.546
#>   std. dev.      0.2621  0.2462
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6773  0.5634
#>   std. dev.       0.246   0.259
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6868  0.6173
#>   std. dev.      0.2596  0.2521
#>   weight sum         77      62
#>   precision      0.0071  0.0071
#> 
#> V24
#>   mean           0.6993  0.6605
#>   std. dev.      0.2468  0.2424
#>   weight sum         77      62
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6869  0.6561
#>   std. dev.      0.2469  0.2555
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V26
#>   mean           0.7014   0.675
#>   std. dev.      0.2349  0.2458
#>   weight sum         77      62
#>   precision      0.0063  0.0063
#> 
#> V27
#>   mean           0.7081  0.6691
#>   std. dev.      0.2601  0.2246
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V28
#>   mean           0.7172  0.6524
#>   std. dev.      0.2481  0.2197
#>   weight sum         77      62
#>   precision      0.0075  0.0075
#> 
#> V29
#>   mean            0.647  0.6227
#>   std. dev.      0.2434  0.2378
#>   weight sum         77      62
#>   precision      0.0076  0.0076
#> 
#> V3
#>   mean           0.0525  0.0373
#>   std. dev.      0.0478  0.0292
#>   weight sum         77      62
#>   precision      0.0024  0.0024
#> 
#> V30
#>   mean           0.5774  0.5713
#>   std. dev.      0.2052  0.2335
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4582  0.5264
#>   std. dev.      0.2161  0.2073
#>   weight sum         77      62
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4067  0.4475
#>   std. dev.      0.1988  0.2159
#>   weight sum         77      62
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.3947  0.4419
#>   std. dev.      0.1841  0.2263
#>   weight sum         77      62
#>   precision      0.0068  0.0068
#> 
#> V34
#>   mean           0.3714  0.4448
#>   std. dev.      0.2054  0.2573
#>   weight sum         77      62
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean            0.351  0.4579
#>   std. dev.      0.2467  0.2708
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3339  0.4598
#>   std. dev.      0.2711  0.2767
#>   weight sum         77      62
#>   precision      0.0073  0.0073
#> 
#> V37
#>   mean           0.3324  0.4108
#>   std. dev.      0.2504  0.2514
#>   weight sum         77      62
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean            0.342  0.3476
#>   std. dev.      0.2262  0.2296
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3402  0.3211
#>   std. dev.      0.1949  0.2168
#>   weight sum         77      62
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0699   0.042
#>   std. dev.      0.0598  0.0325
#>   weight sum         77      62
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean            0.311  0.3257
#>   std. dev.      0.1657  0.2004
#>   weight sum         77      62
#>   precision      0.0066  0.0066
#> 
#> V41
#>   mean           0.2884  0.2956
#>   std. dev.      0.1506  0.1746
#>   weight sum         77      62
#>   precision      0.0054  0.0054
#> 
#> V42
#>   mean           0.2831  0.2626
#>   std. dev.      0.1462   0.164
#>   weight sum         77      62
#>   precision       0.005   0.005
#> 
#> V43
#>   mean           0.2641  0.2255
#>   std. dev.      0.1235  0.1323
#>   weight sum         77      62
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2363  0.1881
#>   std. dev.      0.1312  0.1102
#>   weight sum         77      62
#>   precision      0.0057  0.0057
#> 
#> V45
#>   mean            0.234   0.146
#>   std. dev.      0.1556  0.0959
#>   weight sum         77      62
#>   precision      0.0042  0.0042
#> 
#> V46
#>   mean           0.1893  0.1158
#>   std. dev.      0.1314  0.0867
#>   weight sum         77      62
#>   precision      0.0042  0.0042
#> 
#> V47
#>   mean           0.1365  0.0953
#>   std. dev.      0.0789  0.0592
#>   weight sum         77      62
#>   precision      0.0028  0.0028
#> 
#> V48
#>   mean           0.1065  0.0719
#>   std. dev.       0.061  0.0437
#>   weight sum         77      62
#>   precision      0.0021  0.0021
#> 
#> V49
#>   mean           0.0625  0.0405
#>   std. dev.      0.0319  0.0259
#>   weight sum         77      62
#>   precision      0.0011  0.0011
#> 
#> V5
#>   mean           0.0944  0.0618
#>   std. dev.      0.0658  0.0473
#>   weight sum         77      62
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0218  0.0176
#>   std. dev.      0.0138  0.0112
#>   weight sum         77      62
#>   precision      0.0008  0.0008
#> 
#> V51
#>   mean           0.0175  0.0122
#>   std. dev.      0.0087  0.0076
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V52
#>   mean           0.0159  0.0106
#>   std. dev.      0.0091  0.0071
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0118  0.0099
#>   std. dev.      0.0077  0.0057
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0123  0.0091
#>   std. dev.      0.0085  0.0048
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V55
#>   mean           0.0097  0.0087
#>   std. dev.       0.008  0.0052
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0094  0.0075
#>   std. dev.      0.0069   0.005
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0082  0.0084
#>   std. dev.      0.0064  0.0063
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean            0.009  0.0073
#>   std. dev.      0.0073  0.0051
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0088  0.0073
#>   std. dev.      0.0065  0.0054
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1154  0.0967
#>   std. dev.      0.0543  0.0692
#>   weight sum         77      62
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean           0.0072  0.0059
#>   std. dev.      0.0062  0.0034
#>   weight sum         77      62
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1287  0.1174
#>   std. dev.      0.0588  0.0685
#>   weight sum         77      62
#>   precision      0.0028  0.0028
#> 
#> V8
#>   mean           0.1539  0.1182
#>   std. dev.      0.0945  0.0886
#>   weight sum         77      62
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2162   0.138
#>   std. dev.      0.1349  0.1077
#>   weight sum         77      62
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696