Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.51)  (0.49)
#> ===============================
#> V1
#>   mean           0.0359  0.0229
#>   std. dev.      0.0293  0.0139
#>   weight sum         71      68
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2553  0.1668
#>   std. dev.      0.1347  0.1262
#>   weight sum         71      68
#>   precision      0.0047  0.0047
#> 
#> V11
#>   mean           0.2941  0.1813
#>   std. dev.      0.1248   0.127
#>   weight sum         71      68
#>   precision      0.0047  0.0047
#> 
#> V12
#>   mean           0.3083  0.1974
#>   std. dev.      0.1233  0.1444
#>   weight sum         71      68
#>   precision      0.0049  0.0049
#> 
#> V13
#>   mean           0.3173  0.2308
#>   std. dev.      0.1382  0.1395
#>   weight sum         71      68
#>   precision      0.0053  0.0053
#> 
#> V14
#>   mean           0.3244  0.2704
#>   std. dev.      0.1394  0.1722
#>   weight sum         71      68
#>   precision      0.0055  0.0055
#> 
#> V15
#>   mean           0.3376   0.319
#>   std. dev.      0.1793  0.2178
#>   weight sum         71      68
#>   precision      0.0061  0.0061
#> 
#> V16
#>   mean           0.3879  0.3958
#>   std. dev.      0.2125  0.2634
#>   weight sum         71      68
#>   precision       0.007   0.007
#> 
#> V17
#>   mean           0.4234  0.4413
#>   std. dev.      0.2484  0.2989
#>   weight sum         71      68
#>   precision      0.0069  0.0069
#> 
#> V18
#>   mean           0.4545  0.4717
#>   std. dev.      0.2582  0.2803
#>   weight sum         71      68
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean            0.533  0.4802
#>   std. dev.      0.2465  0.2654
#>   weight sum         71      68
#>   precision      0.0067  0.0067
#> 
#> V2
#>   mean           0.0463  0.0319
#>   std. dev.      0.0397   0.026
#>   weight sum         71      68
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6019  0.5081
#>   std. dev.      0.2587  0.2603
#>   weight sum         71      68
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6466  0.5509
#>   std. dev.      0.2612  0.2543
#>   weight sum         71      68
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6423  0.5698
#>   std. dev.      0.2529  0.2693
#>   weight sum         71      68
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6355  0.5953
#>   std. dev.      0.2568  0.2471
#>   weight sum         71      68
#>   precision      0.0071  0.0071
#> 
#> V24
#>   mean           0.6539  0.6392
#>   std. dev.      0.2463  0.2373
#>   weight sum         71      68
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6522  0.6464
#>   std. dev.      0.2427   0.262
#>   weight sum         71      68
#>   precision      0.0075  0.0075
#> 
#> V26
#>   mean            0.685  0.6543
#>   std. dev.      0.2554  0.2472
#>   weight sum         71      68
#>   precision      0.0068  0.0068
#> 
#> V27
#>   mean           0.7165  0.6552
#>   std. dev.      0.2832   0.219
#>   weight sum         71      68
#>   precision      0.0076  0.0076
#> 
#> V28
#>   mean           0.7337  0.6601
#>   std. dev.      0.2662  0.1979
#>   weight sum         71      68
#>   precision      0.0075  0.0075
#> 
#> V29
#>   mean           0.6749  0.6337
#>   std. dev.      0.2502  0.2306
#>   weight sum         71      68
#>   precision      0.0072  0.0072
#> 
#> V3
#>   mean           0.0491  0.0367
#>   std. dev.      0.0439   0.031
#>   weight sum         71      68
#>   precision      0.0024  0.0024
#> 
#> V30
#>   mean           0.5992  0.6078
#>   std. dev.      0.2257  0.2224
#>   weight sum         71      68
#>   precision      0.0065  0.0065
#> 
#> V31
#>   mean           0.4932  0.5709
#>   std. dev.      0.2277  0.1886
#>   weight sum         71      68
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean            0.452  0.4876
#>   std. dev.      0.2006  0.2056
#>   weight sum         71      68
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4145  0.4619
#>   std. dev.      0.1843  0.2127
#>   weight sum         71      68
#>   precision      0.0066  0.0066
#> 
#> V34
#>   mean           0.3722  0.4583
#>   std. dev.      0.2033  0.2577
#>   weight sum         71      68
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3403  0.4658
#>   std. dev.      0.2436  0.2757
#>   weight sum         71      68
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3201  0.4737
#>   std. dev.      0.2406  0.2737
#>   weight sum         71      68
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3284  0.4263
#>   std. dev.      0.2226  0.2379
#>   weight sum         71      68
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean            0.344   0.361
#>   std. dev.      0.1966  0.2278
#>   weight sum         71      68
#>   precision      0.0069  0.0069
#> 
#> V39
#>   mean           0.3395  0.3375
#>   std. dev.      0.1822   0.232
#>   weight sum         71      68
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0636  0.0409
#>   std. dev.      0.0573  0.0337
#>   weight sum         71      68
#>   precision      0.0034  0.0034
#> 
#> V40
#>   mean           0.3111  0.3487
#>   std. dev.      0.1675  0.2133
#>   weight sum         71      68
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.3009  0.2967
#>   std. dev.      0.1696  0.1931
#>   weight sum         71      68
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.3059  0.2557
#>   std. dev.      0.1711   0.173
#>   weight sum         71      68
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2759  0.2173
#>   std. dev.      0.1516  0.1374
#>   weight sum         71      68
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2508  0.1898
#>   std. dev.      0.1535  0.1131
#>   weight sum         71      68
#>   precision      0.0057  0.0057
#> 
#> V45
#>   mean           0.2444   0.156
#>   std. dev.      0.1758  0.1041
#>   weight sum         71      68
#>   precision      0.0046  0.0046
#> 
#> V46
#>   mean           0.1982  0.1259
#>   std. dev.      0.1513  0.1019
#>   weight sum         71      68
#>   precision      0.0055  0.0055
#> 
#> V47
#>   mean           0.1426  0.1036
#>   std. dev.      0.0965  0.0711
#>   weight sum         71      68
#>   precision       0.004   0.004
#> 
#> V48
#>   mean           0.1076  0.0746
#>   std. dev.      0.0667  0.0493
#>   weight sum         71      68
#>   precision      0.0024  0.0024
#> 
#> V49
#>   mean           0.0613  0.0403
#>   std. dev.      0.0369  0.0295
#>   weight sum         71      68
#>   precision      0.0014  0.0014
#> 
#> V5
#>   mean           0.0885   0.066
#>   std. dev.      0.0579  0.0527
#>   weight sum         71      68
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0223  0.0184
#>   std. dev.      0.0136  0.0132
#>   weight sum         71      68
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean            0.019  0.0124
#>   std. dev.       0.015  0.0084
#>   weight sum         71      68
#>   precision      0.0009  0.0009
#> 
#> V52
#>   mean           0.0163    0.01
#>   std. dev.      0.0113  0.0065
#>   weight sum         71      68
#>   precision      0.0006  0.0006
#> 
#> V53
#>   mean           0.0117  0.0097
#>   std. dev.      0.0075  0.0061
#>   weight sum         71      68
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0119  0.0099
#>   std. dev.      0.0076  0.0051
#>   weight sum         71      68
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0093   0.009
#>   std. dev.      0.0066  0.0052
#>   weight sum         71      68
#>   precision      0.0003  0.0003
#> 
#> V56
#>   mean           0.0087  0.0076
#>   std. dev.      0.0049  0.0049
#>   weight sum         71      68
#>   precision      0.0003  0.0003
#> 
#> V57
#>   mean           0.0072  0.0075
#>   std. dev.       0.005  0.0057
#>   weight sum         71      68
#>   precision      0.0003  0.0003
#> 
#> V58
#>   mean           0.0086  0.0065
#>   std. dev.      0.0064  0.0048
#>   weight sum         71      68
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0083  0.0077
#>   std. dev.      0.0061  0.0055
#>   weight sum         71      68
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1113  0.1013
#>   std. dev.      0.0522  0.0728
#>   weight sum         71      68
#>   precision      0.0029  0.0029
#> 
#> V60
#>   mean           0.0071  0.0063
#>   std. dev.      0.0065  0.0038
#>   weight sum         71      68
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1296  0.1219
#>   std. dev.      0.0554  0.0688
#>   weight sum         71      68
#>   precision      0.0027  0.0027
#> 
#> V8
#>   mean           0.1446  0.1251
#>   std. dev.      0.0829  0.0839
#>   weight sum         71      68
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2163  0.1451
#>   std. dev.      0.1188  0.1132
#>   weight sum         71      68
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913