Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$marshal()

LearnerClassifNaiveBayesWeka$unmarshal()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0356  0.0225
#>   std. dev.      0.0274  0.0127
#>   weight sum         73      66
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2525   0.164
#>   std. dev.      0.1259  0.1227
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V11
#>   mean           0.2906  0.1842
#>   std. dev.      0.1108   0.125
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V12
#>   mean           0.3001   0.204
#>   std. dev.      0.1142  0.1464
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3119  0.2356
#>   std. dev.      0.1262  0.1412
#>   weight sum         73      66
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3141  0.2729
#>   std. dev.       0.164  0.1714
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3213  0.3131
#>   std. dev.      0.1928  0.2194
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V16
#>   mean           0.3735   0.383
#>   std. dev.      0.2058   0.256
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V17
#>   mean           0.4129  0.4279
#>   std. dev.      0.2274  0.2972
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4533  0.4573
#>   std. dev.      0.2438  0.2809
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5377  0.4828
#>   std. dev.      0.2533  0.2666
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0441  0.0312
#>   std. dev.      0.0355  0.0257
#>   weight sum         73      66
#>   precision      0.0013  0.0013
#> 
#> V20
#>   mean           0.6101  0.5196
#>   std. dev.      0.2608   0.278
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6638   0.548
#>   std. dev.       0.264  0.2682
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6838  0.5573
#>   std. dev.      0.2433  0.2667
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6922  0.5966
#>   std. dev.      0.2482  0.2418
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V24
#>   mean           0.7008  0.6481
#>   std. dev.      0.2285  0.2278
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V25
#>   mean           0.6983  0.6648
#>   std. dev.      0.2281  0.2525
#>   weight sum         73      66
#>   precision      0.0074  0.0074
#> 
#> V26
#>   mean           0.7253  0.7012
#>   std. dev.      0.2267  0.2455
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V27
#>   mean           0.7301  0.6875
#>   std. dev.      0.2617  0.2187
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V28
#>   mean           0.7234  0.6644
#>   std. dev.        0.26   0.201
#>   weight sum         73      66
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6651   0.622
#>   std. dev.      0.2412  0.2337
#>   weight sum         73      66
#>   precision      0.0074  0.0074
#> 
#> V3
#>   mean           0.0447  0.0361
#>   std. dev.      0.0338  0.0285
#>   weight sum         73      66
#>   precision      0.0013  0.0013
#> 
#> V30
#>   mean           0.6003  0.5721
#>   std. dev.      0.2076  0.2365
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.5049  0.5219
#>   std. dev.      0.2334  0.1995
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4624  0.4332
#>   std. dev.      0.2224  0.2122
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4176  0.4299
#>   std. dev.      0.2062  0.2186
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3784  0.4545
#>   std. dev.      0.2189  0.2547
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean            0.357   0.454
#>   std. dev.      0.2611  0.2659
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3416  0.4561
#>   std. dev.      0.2529  0.2734
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V37
#>   mean           0.3341  0.4139
#>   std. dev.       0.234  0.2508
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean            0.342  0.3371
#>   std. dev.      0.2029  0.2216
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V39
#>   mean           0.3534  0.2974
#>   std. dev.      0.1853  0.2071
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V4
#>   mean             0.06  0.0436
#>   std. dev.      0.0392  0.0326
#>   weight sum         73      66
#>   precision      0.0013  0.0013
#> 
#> V40
#>   mean           0.3177  0.3151
#>   std. dev.      0.1615   0.185
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V41
#>   mean           0.3044  0.2959
#>   std. dev.      0.1737  0.1723
#>   weight sum         73      66
#>   precision      0.0064  0.0064
#> 
#> V42
#>   mean           0.3219  0.2598
#>   std. dev.      0.1782  0.1683
#>   weight sum         73      66
#>   precision      0.0058  0.0058
#> 
#> V43
#>   mean           0.2941  0.2066
#>   std. dev.      0.1434  0.1323
#>   weight sum         73      66
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2638  0.1664
#>   std. dev.      0.1516  0.1086
#>   weight sum         73      66
#>   precision      0.0056  0.0056
#> 
#> V45
#>   mean           0.2599  0.1411
#>   std. dev.      0.1816  0.0973
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V46
#>   mean           0.2142  0.1189
#>   std. dev.      0.1574  0.0973
#>   weight sum         73      66
#>   precision      0.0053  0.0053
#> 
#> V47
#>   mean           0.1542  0.0928
#>   std. dev.      0.0952  0.0689
#>   weight sum         73      66
#>   precision      0.0041  0.0041
#> 
#> V48
#>   mean           0.1115   0.068
#>   std. dev.      0.0697  0.0484
#>   weight sum         73      66
#>   precision      0.0025  0.0025
#> 
#> V49
#>   mean           0.0651  0.0385
#>   std. dev.      0.0378  0.0296
#>   weight sum         73      66
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean            0.086  0.0667
#>   std. dev.      0.0493  0.0487
#>   weight sum         73      66
#>   precision      0.0019  0.0019
#> 
#> V50
#>   mean           0.0242  0.0168
#>   std. dev.      0.0152  0.0124
#>   weight sum         73      66
#>   precision      0.0008  0.0008
#> 
#> V51
#>   mean           0.0199  0.0127
#>   std. dev.      0.0153  0.0092
#>   weight sum         73      66
#>   precision      0.0008  0.0008
#> 
#> V52
#>   mean           0.0163  0.0109
#>   std. dev.      0.0117  0.0077
#>   weight sum         73      66
#>   precision      0.0006  0.0006
#> 
#> V53
#>   mean           0.0113  0.0098
#>   std. dev.      0.0078  0.0056
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0123   0.009
#>   std. dev.      0.0083  0.0053
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0098   0.009
#>   std. dev.       0.009  0.0054
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0085  0.0074
#>   std. dev.       0.006   0.005
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0074  0.0077
#>   std. dev.      0.0058  0.0056
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean            0.009  0.0066
#>   std. dev.      0.0076  0.0051
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0089  0.0067
#>   std. dev.      0.0074  0.0054
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1159  0.1029
#>   std. dev.      0.0532  0.0693
#>   weight sum         73      66
#>   precision      0.0029  0.0029
#> 
#> V60
#>   mean           0.0073  0.0058
#>   std. dev.      0.0064  0.0036
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1373   0.119
#>   std. dev.      0.0586   0.069
#>   weight sum         73      66
#>   precision      0.0027  0.0027
#> 
#> V8
#>   mean           0.1475  0.1233
#>   std. dev.      0.0826  0.0865
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2123  0.1429
#>   std. dev.      0.1146  0.1069
#>   weight sum         73      66
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2753623