Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0339  0.0231
#>   std. dev.      0.0246  0.0151
#>   weight sum         72      67
#>   precision      0.0009  0.0009
#> 
#> V10
#>   mean           0.2492   0.164
#>   std. dev.      0.1325  0.1217
#>   weight sum         72      67
#>   precision       0.005   0.005
#> 
#> V11
#>   mean           0.2851  0.1765
#>   std. dev.      0.1153  0.1188
#>   weight sum         72      67
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.3002  0.1907
#>   std. dev.      0.1113  0.1356
#>   weight sum         72      67
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3093  0.2283
#>   std. dev.      0.1191  0.1296
#>   weight sum         72      67
#>   precision      0.0051  0.0051
#> 
#> V14
#>   mean           0.3178  0.2653
#>   std. dev.      0.1637  0.1577
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V15
#>   mean           0.3297  0.3029
#>   std. dev.      0.1935  0.2151
#>   weight sum         72      67
#>   precision      0.0073  0.0073
#> 
#> V16
#>   mean           0.3701   0.373
#>   std. dev.      0.2017  0.2594
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V17
#>   mean           0.3964  0.4103
#>   std. dev.      0.2218  0.2968
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V18
#>   mean           0.4431  0.4446
#>   std. dev.      0.2489  0.2757
#>   weight sum         72      67
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5364  0.4748
#>   std. dev.      0.2559  0.2627
#>   weight sum         72      67
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0448   0.031
#>   std. dev.      0.0333  0.0249
#>   weight sum         72      67
#>   precision      0.0013  0.0013
#> 
#> V20
#>   mean           0.6198  0.5119
#>   std. dev.      0.2563  0.2723
#>   weight sum         72      67
#>   precision      0.0068  0.0068
#> 
#> V21
#>   mean           0.6675  0.5427
#>   std. dev.      0.2566  0.2614
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6722  0.5506
#>   std. dev.      0.2457  0.2703
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6754  0.6053
#>   std. dev.      0.2517  0.2471
#>   weight sum         72      67
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6944  0.6525
#>   std. dev.      0.2305  0.2456
#>   weight sum         72      67
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6917  0.6585
#>   std. dev.      0.2214  0.2666
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V26
#>   mean           0.7259  0.6711
#>   std. dev.      0.2162  0.2527
#>   weight sum         72      67
#>   precision      0.0066  0.0066
#> 
#> V27
#>   mean           0.7398  0.6857
#>   std. dev.      0.2558  0.2236
#>   weight sum         72      67
#>   precision      0.0071  0.0071
#> 
#> V28
#>   mean           0.7272  0.6946
#>   std. dev.      0.2586  0.1914
#>   weight sum         72      67
#>   precision      0.0075  0.0075
#> 
#> V29
#>   mean           0.6733  0.6306
#>   std. dev.      0.2317  0.2389
#>   weight sum         72      67
#>   precision      0.0071  0.0071
#> 
#> V3
#>   mean           0.0479  0.0357
#>   std. dev.      0.0364  0.0285
#>   weight sum         72      67
#>   precision      0.0015  0.0015
#> 
#> V30
#>   mean           0.6058  0.5675
#>   std. dev.      0.2027  0.2275
#>   weight sum         72      67
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4964  0.5241
#>   std. dev.      0.2199  0.1941
#>   weight sum         72      67
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4236  0.4539
#>   std. dev.      0.2189  0.2152
#>   weight sum         72      67
#>   precision      0.0063  0.0063
#> 
#> V33
#>   mean           0.3896  0.4472
#>   std. dev.      0.2058  0.2198
#>   weight sum         72      67
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3603   0.443
#>   std. dev.      0.2129  0.2542
#>   weight sum         72      67
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean           0.3367   0.451
#>   std. dev.      0.2523  0.2599
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3184  0.4615
#>   std. dev.      0.2597  0.2641
#>   weight sum         72      67
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean             0.33  0.4246
#>   std. dev.      0.2378  0.2424
#>   weight sum         72      67
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3352  0.3489
#>   std. dev.       0.209  0.2129
#>   weight sum         72      67
#>   precision      0.0065  0.0065
#> 
#> V39
#>   mean           0.3364  0.3091
#>   std. dev.       0.186  0.2005
#>   weight sum         72      67
#>   precision       0.007   0.007
#> 
#> V4
#>   mean           0.0622  0.0437
#>   std. dev.      0.0429  0.0322
#>   weight sum         72      67
#>   precision       0.002   0.002
#> 
#> V40
#>   mean           0.3144  0.3104
#>   std. dev.      0.1606  0.1789
#>   weight sum         72      67
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2968   0.275
#>   std. dev.      0.1661  0.1618
#>   weight sum         72      67
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.3024  0.2489
#>   std. dev.      0.1506  0.1564
#>   weight sum         72      67
#>   precision      0.0057  0.0057
#> 
#> V43
#>   mean           0.2787  0.2008
#>   std. dev.      0.1247  0.1109
#>   weight sum         72      67
#>   precision      0.0043  0.0043
#> 
#> V44
#>   mean           0.2406  0.1639
#>   std. dev.      0.1423  0.0845
#>   weight sum         72      67
#>   precision      0.0043  0.0043
#> 
#> V45
#>   mean           0.2425  0.1352
#>   std. dev.      0.1717  0.0818
#>   weight sum         72      67
#>   precision      0.0051  0.0051
#> 
#> V46
#>   mean           0.1966  0.1108
#>   std. dev.      0.1436  0.0849
#>   weight sum         72      67
#>   precision      0.0044  0.0044
#> 
#> V47
#>   mean           0.1429  0.0891
#>   std. dev.      0.0915  0.0649
#>   weight sum         72      67
#>   precision      0.0032  0.0032
#> 
#> V48
#>   mean           0.1072  0.0672
#>   std. dev.       0.068  0.0493
#>   weight sum         72      67
#>   precision      0.0021  0.0021
#> 
#> V49
#>   mean           0.0645  0.0388
#>   std. dev.      0.0378  0.0334
#>   weight sum         72      67
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0846  0.0621
#>   std. dev.      0.0576  0.0485
#>   weight sum         72      67
#>   precision      0.0024  0.0024
#> 
#> V50
#>   mean            0.023  0.0167
#>   std. dev.      0.0155  0.0132
#>   weight sum         72      67
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0182  0.0119
#>   std. dev.      0.0114  0.0084
#>   weight sum         72      67
#>   precision      0.0007  0.0007
#> 
#> V52
#>   mean           0.0151  0.0108
#>   std. dev.       0.009  0.0072
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean            0.011  0.0102
#>   std. dev.      0.0066  0.0062
#>   weight sum         72      67
#>   precision      0.0003  0.0003
#> 
#> V54
#>   mean            0.012    0.01
#>   std. dev.      0.0081  0.0055
#>   weight sum         72      67
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0103  0.0091
#>   std. dev.      0.0094  0.0054
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0088  0.0075
#>   std. dev.      0.0064  0.0046
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0077  0.0079
#>   std. dev.       0.006  0.0059
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0098  0.0066
#>   std. dev.      0.0085  0.0047
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0088  0.0075
#>   std. dev.      0.0069  0.0051
#>   weight sum         72      67
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1184  0.0949
#>   std. dev.      0.0512  0.0623
#>   weight sum         72      67
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0069   0.006
#>   std. dev.      0.0063  0.0036
#>   weight sum         72      67
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1332  0.1152
#>   std. dev.      0.0598  0.0648
#>   weight sum         72      67
#>   precision      0.0025  0.0025
#> 
#> V8
#>   mean           0.1542  0.1163
#>   std. dev.      0.0867  0.0806
#>   weight sum         72      67
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2154  0.1384
#>   std. dev.       0.113  0.1043
#>   weight sum         72      67
#>   precision       0.005   0.005
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406