Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0348  0.0239
#>   std. dev.      0.0261  0.0159
#>   weight sum         73      66
#>   precision       0.001   0.001
#> 
#> V10
#>   mean           0.2424  0.1635
#>   std. dev.      0.1198  0.1114
#>   weight sum         73      66
#>   precision      0.0044  0.0044
#> 
#> V11
#>   mean           0.2798  0.1791
#>   std. dev.       0.118  0.1089
#>   weight sum         73      66
#>   precision      0.0045  0.0045
#> 
#> V12
#>   mean           0.2987  0.2016
#>   std. dev.      0.1212  0.1396
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V13
#>   mean            0.308  0.2371
#>   std. dev.      0.1291  0.1419
#>   weight sum         73      66
#>   precision      0.0048  0.0048
#> 
#> V14
#>   mean           0.3153  0.2752
#>   std. dev.      0.1544  0.1728
#>   weight sum         73      66
#>   precision      0.0058  0.0058
#> 
#> V15
#>   mean           0.3284  0.3193
#>   std. dev.      0.1975  0.2242
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V16
#>   mean           0.3776  0.3926
#>   std. dev.      0.2203  0.2566
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V17
#>   mean             0.43  0.4361
#>   std. dev.      0.2454   0.281
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V18
#>   mean           0.4764  0.4557
#>   std. dev.      0.2502  0.2677
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V19
#>   mean           0.5477  0.4578
#>   std. dev.      0.2502  0.2653
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0452  0.0304
#>   std. dev.      0.0389  0.0213
#>   weight sum         73      66
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6165  0.4947
#>   std. dev.       0.256    0.26
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V21
#>   mean           0.6576  0.5339
#>   std. dev.      0.2493  0.2445
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V22
#>   mean           0.6774  0.5569
#>   std. dev.      0.2335  0.2533
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6796  0.6105
#>   std. dev.      0.2498  0.2449
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V24
#>   mean           0.6905  0.6564
#>   std. dev.      0.2436  0.2284
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V25
#>   mean           0.6912  0.6848
#>   std. dev.      0.2452  0.2394
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean            0.715  0.7027
#>   std. dev.      0.2412  0.2434
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V27
#>   mean           0.7247  0.6859
#>   std. dev.      0.2678  0.2265
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V28
#>   mean           0.7068   0.672
#>   std. dev.      0.2635  0.2072
#>   weight sum         73      66
#>   precision      0.0077  0.0077
#> 
#> V29
#>   mean           0.6418  0.6179
#>   std. dev.      0.2413  0.2259
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V3
#>   mean           0.0538  0.0357
#>   std. dev.      0.0495  0.0266
#>   weight sum         73      66
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5763  0.5531
#>   std. dev.      0.2095  0.2302
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V31
#>   mean             0.48  0.5175
#>   std. dev.      0.2196   0.201
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4275  0.4654
#>   std. dev.      0.2154   0.206
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.3933  0.4422
#>   std. dev.      0.1959  0.2192
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V34
#>   mean           0.3434  0.4575
#>   std. dev.      0.1975  0.2519
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3205  0.4796
#>   std. dev.      0.2433  0.2643
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3105  0.4891
#>   std. dev.      0.2478   0.264
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3305  0.4339
#>   std. dev.      0.2275  0.2524
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean            0.334  0.3503
#>   std. dev.      0.2024  0.2282
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3263  0.3153
#>   std. dev.      0.1805  0.1931
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V4
#>   mean           0.0679  0.0431
#>   std. dev.      0.0619  0.0277
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.2928  0.3142
#>   std. dev.      0.1658  0.1779
#>   weight sum         73      66
#>   precision      0.0058  0.0058
#> 
#> V41
#>   mean           0.2799  0.2833
#>   std. dev.      0.1656  0.1662
#>   weight sum         73      66
#>   precision      0.0054  0.0054
#> 
#> V42
#>   mean           0.2796   0.246
#>   std. dev.      0.1589   0.152
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V43
#>   mean           0.2581  0.2063
#>   std. dev.      0.1272  0.1107
#>   weight sum         73      66
#>   precision      0.0044  0.0044
#> 
#> V44
#>   mean           0.2372  0.1699
#>   std. dev.      0.1428  0.0884
#>   weight sum         73      66
#>   precision      0.0042  0.0042
#> 
#> V45
#>   mean           0.2235  0.1302
#>   std. dev.      0.1685  0.0678
#>   weight sum         73      66
#>   precision      0.0045  0.0045
#> 
#> V46
#>   mean            0.181  0.1091
#>   std. dev.      0.1363  0.0703
#>   weight sum         73      66
#>   precision      0.0042  0.0042
#> 
#> V47
#>   mean           0.1378  0.0846
#>   std. dev.      0.0814  0.0509
#>   weight sum         73      66
#>   precision      0.0027  0.0027
#> 
#> V48
#>   mean           0.1012   0.062
#>   std. dev.        0.06  0.0376
#>   weight sum         73      66
#>   precision      0.0021  0.0021
#> 
#> V49
#>   mean           0.0572   0.035
#>   std. dev.      0.0316  0.0249
#>   weight sum         73      66
#>   precision      0.0011  0.0011
#> 
#> V5
#>   mean            0.092  0.0591
#>   std. dev.      0.0678  0.0429
#>   weight sum         73      66
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0214  0.0179
#>   std. dev.      0.0126  0.0119
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0172  0.0125
#>   std. dev.      0.0089  0.0082
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V52
#>   mean           0.0157  0.0114
#>   std. dev.      0.0094  0.0077
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0116  0.0097
#>   std. dev.      0.0072  0.0061
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V54
#>   mean           0.0115  0.0093
#>   std. dev.      0.0072  0.0056
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0093  0.0084
#>   std. dev.      0.0084  0.0051
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0089  0.0076
#>   std. dev.      0.0064  0.0049
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0082  0.0084
#>   std. dev.       0.006   0.006
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0088  0.0068
#>   std. dev.      0.0078   0.005
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0087  0.0068
#>   std. dev.      0.0071   0.005
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1107  0.0981
#>   std. dev.      0.0519  0.0707
#>   weight sum         73      66
#>   precision      0.0029  0.0029
#> 
#> V60
#>   mean           0.0076  0.0054
#>   std. dev.      0.0066  0.0029
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1253  0.1147
#>   std. dev.      0.0599   0.065
#>   weight sum         73      66
#>   precision      0.0028  0.0028
#> 
#> V8
#>   mean           0.1354  0.1208
#>   std. dev.      0.0809  0.0777
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean            0.206  0.1455
#>   std. dev.      0.1119  0.1012
#>   weight sum         73      66
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551