Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.53)  (0.47)
#> ===============================
#> V1
#>   mean           0.0364  0.0245
#>   std. dev.      0.0277   0.016
#>   weight sum         74      65
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2616  0.1672
#>   std. dev.      0.1527  0.1151
#>   weight sum         74      65
#>   precision       0.005   0.005
#> 
#> V11
#>   mean           0.2954  0.1799
#>   std. dev.      0.1327  0.1134
#>   weight sum         74      65
#>   precision      0.0051  0.0051
#> 
#> V12
#>   mean           0.2921  0.1968
#>   std. dev.      0.1196    0.13
#>   weight sum         74      65
#>   precision      0.0046  0.0046
#> 
#> V13
#>   mean           0.3029  0.2246
#>   std. dev.      0.1243  0.1298
#>   weight sum         74      65
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3171  0.2707
#>   std. dev.      0.1584  0.1558
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3367  0.3135
#>   std. dev.      0.1957  0.2132
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V16
#>   mean           0.3713  0.3934
#>   std. dev.      0.2111  0.2544
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V17
#>   mean            0.386  0.4381
#>   std. dev.      0.2319  0.2969
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4244  0.4622
#>   std. dev.       0.255  0.2747
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean           0.5061   0.471
#>   std. dev.      0.2569  0.2579
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0496  0.0316
#>   std. dev.      0.0391  0.0256
#>   weight sum         74      65
#>   precision      0.0019  0.0019
#> 
#> V20
#>   mean           0.5955  0.5093
#>   std. dev.      0.2602  0.2585
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6534  0.5551
#>   std. dev.      0.2688  0.2535
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean            0.649   0.579
#>   std. dev.      0.2531  0.2717
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V23
#>   mean           0.6496  0.6222
#>   std. dev.      0.2564   0.252
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6646  0.6399
#>   std. dev.      0.2437  0.2464
#>   weight sum         74      65
#>   precision      0.0075  0.0075
#> 
#> V25
#>   mean           0.6502   0.641
#>   std. dev.      0.2389  0.2673
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.6821  0.6551
#>   std. dev.      0.2337  0.2487
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V27
#>   mean           0.7065  0.6585
#>   std. dev.      0.2629  0.2207
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V28
#>   mean           0.7224   0.649
#>   std. dev.      0.2506  0.2072
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V29
#>   mean           0.6683   0.608
#>   std. dev.      0.2253  0.2364
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V3
#>   mean           0.0538  0.0371
#>   std. dev.      0.0464  0.0316
#>   weight sum         74      65
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5943  0.5616
#>   std. dev.        0.21  0.2358
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4791  0.5176
#>   std. dev.      0.2228  0.2016
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4127  0.4445
#>   std. dev.      0.2126   0.214
#>   weight sum         74      65
#>   precision      0.0064  0.0064
#> 
#> V33
#>   mean           0.3854  0.4539
#>   std. dev.      0.1917  0.2134
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V34
#>   mean           0.3816  0.4633
#>   std. dev.      0.2057  0.2585
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3731  0.4651
#>   std. dev.      0.2531  0.2765
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3405  0.4725
#>   std. dev.      0.2685  0.2736
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3298   0.431
#>   std. dev.       0.247  0.2441
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean           0.3499  0.3585
#>   std. dev.      0.2252  0.2334
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3512  0.3193
#>   std. dev.      0.1892  0.2318
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V4
#>   mean           0.0705  0.0421
#>   std. dev.      0.0604  0.0336
#>   weight sum         74      65
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3153  0.3363
#>   std. dev.      0.1524  0.2103
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2915   0.299
#>   std. dev.      0.1605  0.1933
#>   weight sum         74      65
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.2907  0.2704
#>   std. dev.      0.1749  0.1772
#>   weight sum         74      65
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2672  0.2254
#>   std. dev.      0.1435  0.1256
#>   weight sum         74      65
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2469  0.1846
#>   std. dev.      0.1429  0.0927
#>   weight sum         74      65
#>   precision      0.0043  0.0043
#> 
#> V45
#>   mean           0.2536  0.1492
#>   std. dev.      0.1716  0.0889
#>   weight sum         74      65
#>   precision      0.0047  0.0047
#> 
#> V46
#>   mean           0.2098  0.1235
#>   std. dev.      0.1533   0.097
#>   weight sum         74      65
#>   precision      0.0054  0.0054
#> 
#> V47
#>   mean           0.1497  0.1019
#>   std. dev.       0.093  0.0707
#>   weight sum         74      65
#>   precision      0.0041  0.0041
#> 
#> V48
#>   mean           0.1105  0.0753
#>   std. dev.      0.0622  0.0517
#>   weight sum         74      65
#>   precision      0.0024  0.0024
#> 
#> V49
#>   mean           0.0638  0.0426
#>   std. dev.      0.0329  0.0327
#>   weight sum         74      65
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0842  0.0645
#>   std. dev.      0.0623  0.0488
#>   weight sum         74      65
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0219  0.0195
#>   std. dev.      0.0124  0.0134
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0198  0.0134
#>   std. dev.      0.0132  0.0095
#>   weight sum         74      65
#>   precision      0.0008  0.0008
#> 
#> V52
#>   mean           0.0164  0.0107
#>   std. dev.      0.0114  0.0077
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0115  0.0096
#>   std. dev.      0.0074  0.0059
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0124  0.0105
#>   std. dev.      0.0089  0.0053
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0107  0.0088
#>   std. dev.      0.0082  0.0047
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean            0.009  0.0072
#>   std. dev.      0.0061  0.0046
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean            0.008   0.008
#>   std. dev.      0.0054   0.005
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V58
#>   mean           0.0093  0.0069
#>   std. dev.      0.0071  0.0045
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0093   0.007
#>   std. dev.      0.0075  0.0045
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1115  0.0998
#>   std. dev.      0.0529  0.0597
#>   weight sum         74      65
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0076  0.0063
#>   std. dev.      0.0068  0.0035
#>   weight sum         74      65
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1328  0.1172
#>   std. dev.      0.0631  0.0605
#>   weight sum         74      65
#>   precision      0.0025  0.0025
#> 
#> V8
#>   mean            0.162  0.1213
#>   std. dev.      0.0984  0.0769
#>   weight sum         74      65
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean            0.226  0.1399
#>   std. dev.      0.1372  0.0991
#>   weight sum         74      65
#>   precision       0.005   0.005
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3333333