Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                  (0.5)   (0.5)
#> ===============================
#> V1
#>   mean            0.035  0.0225
#>   std. dev.      0.0272  0.0139
#>   weight sum         69      70
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2566  0.1622
#>   std. dev.      0.1393  0.1189
#>   weight sum         69      70
#>   precision       0.005   0.005
#> 
#> V11
#>   mean           0.2908  0.1722
#>   std. dev.      0.1341  0.1226
#>   weight sum         69      70
#>   precision      0.0051  0.0051
#> 
#> V12
#>   mean           0.2942  0.1846
#>   std. dev.       0.134  0.1362
#>   weight sum         69      70
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3076  0.2201
#>   std. dev.      0.1415  0.1317
#>   weight sum         69      70
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean           0.3221  0.2609
#>   std. dev.      0.1746   0.161
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3336   0.301
#>   std. dev.      0.2017  0.2076
#>   weight sum         69      70
#>   precision      0.0067  0.0067
#> 
#> V16
#>   mean           0.3903  0.3771
#>   std. dev.      0.2231  0.2538
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V17
#>   mean           0.4245  0.4224
#>   std. dev.      0.2531  0.2955
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4657  0.4615
#>   std. dev.      0.2663   0.283
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5404  0.4994
#>   std. dev.      0.2642  0.2724
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0445  0.0302
#>   std. dev.       0.037   0.025
#>   weight sum         69      70
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6072   0.527
#>   std. dev.      0.2668   0.269
#>   weight sum         69      70
#>   precision      0.0068  0.0068
#> 
#> V21
#>   mean             0.66   0.564
#>   std. dev.      0.2733  0.2553
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6596  0.5772
#>   std. dev.      0.2627  0.2718
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6652  0.6066
#>   std. dev.      0.2523  0.2572
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V24
#>   mean           0.6729  0.6393
#>   std. dev.      0.2438  0.2362
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean             0.65  0.6411
#>   std. dev.      0.2442  0.2513
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.6747  0.6812
#>   std. dev.      0.2387  0.2498
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V27
#>   mean           0.6923  0.6967
#>   std. dev.      0.2635  0.2144
#>   weight sum         69      70
#>   precision      0.0075  0.0075
#> 
#> V28
#>   mean           0.7034  0.6709
#>   std. dev.      0.2735  0.2048
#>   weight sum         69      70
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6489  0.6269
#>   std. dev.      0.2521  0.2365
#>   weight sum         69      70
#>   precision      0.0073  0.0073
#> 
#> V3
#>   mean           0.0473  0.0348
#>   std. dev.      0.0436  0.0287
#>   weight sum         69      70
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5887  0.5712
#>   std. dev.      0.2246  0.2466
#>   weight sum         69      70
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4878  0.5174
#>   std. dev.      0.2344  0.2079
#>   weight sum         69      70
#>   precision      0.0063  0.0063
#> 
#> V32
#>   mean           0.4213  0.4344
#>   std. dev.      0.2195  0.2142
#>   weight sum         69      70
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.3866  0.4323
#>   std. dev.      0.1863  0.2183
#>   weight sum         69      70
#>   precision      0.0067  0.0067
#> 
#> V34
#>   mean           0.3626  0.4535
#>   std. dev.      0.1969  0.2542
#>   weight sum         69      70
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean           0.3342  0.4674
#>   std. dev.      0.2284  0.2722
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3106  0.4659
#>   std. dev.      0.2422    0.27
#>   weight sum         69      70
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3097  0.4131
#>   std. dev.      0.2266  0.2409
#>   weight sum         69      70
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3162  0.3464
#>   std. dev.      0.2023  0.2137
#>   weight sum         69      70
#>   precision      0.0071  0.0071
#> 
#> V39
#>   mean           0.3122  0.3118
#>   std. dev.       0.174  0.1919
#>   weight sum         69      70
#>   precision      0.0062  0.0062
#> 
#> V4
#>   mean            0.061  0.0413
#>   std. dev.      0.0554  0.0298
#>   weight sum         69      70
#>   precision      0.0034  0.0034
#> 
#> V40
#>   mean           0.2978  0.3121
#>   std. dev.      0.1551  0.1776
#>   weight sum         69      70
#>   precision      0.0064  0.0064
#> 
#> V41
#>   mean           0.2876  0.2803
#>   std. dev.      0.1704  0.1763
#>   weight sum         69      70
#>   precision      0.0054  0.0054
#> 
#> V42
#>   mean           0.2915  0.2421
#>   std. dev.      0.1678  0.1679
#>   weight sum         69      70
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2799  0.2073
#>   std. dev.      0.1474  0.1341
#>   weight sum         69      70
#>   precision      0.0056  0.0056
#> 
#> V44
#>   mean           0.2551   0.168
#>   std. dev.       0.148  0.1166
#>   weight sum         69      70
#>   precision      0.0058  0.0058
#> 
#> V45
#>   mean           0.2387  0.1331
#>   std. dev.      0.1676  0.0894
#>   weight sum         69      70
#>   precision      0.0052  0.0052
#> 
#> V46
#>   mean           0.1919  0.1029
#>   std. dev.       0.152  0.0828
#>   weight sum         69      70
#>   precision      0.0054  0.0054
#> 
#> V47
#>   mean           0.1503  0.0883
#>   std. dev.      0.1011   0.056
#>   weight sum         69      70
#>   precision      0.0041  0.0041
#> 
#> V48
#>   mean           0.1148  0.0655
#>   std. dev.      0.0697  0.0409
#>   weight sum         69      70
#>   precision      0.0025  0.0025
#> 
#> V49
#>   mean           0.0661   0.036
#>   std. dev.      0.0371  0.0231
#>   weight sum         69      70
#>   precision      0.0013  0.0013
#> 
#> V5
#>   mean           0.0821  0.0643
#>   std. dev.       0.056  0.0452
#>   weight sum         69      70
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0231  0.0172
#>   std. dev.      0.0129  0.0113
#>   weight sum         69      70
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean             0.02  0.0118
#>   std. dev.      0.0149  0.0084
#>   weight sum         69      70
#>   precision      0.0009  0.0009
#> 
#> V52
#>   mean           0.0158  0.0107
#>   std. dev.      0.0109  0.0075
#>   weight sum         69      70
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0111  0.0097
#>   std. dev.      0.0073  0.0063
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V54
#>   mean           0.0114  0.0095
#>   std. dev.      0.0082  0.0052
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0091  0.0086
#>   std. dev.      0.0074  0.0048
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0086  0.0072
#>   std. dev.      0.0056   0.005
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V57
#>   mean           0.0073  0.0079
#>   std. dev.      0.0052  0.0059
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V58
#>   mean           0.0094  0.0066
#>   std. dev.      0.0075  0.0052
#>   weight sum         69      70
#>   precision      0.0004  0.0004
#> 
#> V59
#>   mean           0.0083  0.0073
#>   std. dev.      0.0061   0.005
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V6
#>   mean           0.1098  0.0983
#>   std. dev.      0.0518  0.0636
#>   weight sum         69      70
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean           0.0068  0.0057
#>   std. dev.      0.0048  0.0036
#>   weight sum         69      70
#>   precision      0.0003  0.0003
#> 
#> V7
#>   mean           0.1294  0.1111
#>   std. dev.      0.0525  0.0651
#>   weight sum         69      70
#>   precision      0.0028  0.0028
#> 
#> V8
#>   mean           0.1543   0.113
#>   std. dev.      0.0886  0.0838
#>   weight sum         69      70
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2174   0.141
#>   std. dev.      0.1258  0.1061
#>   weight sum         69      70
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3188406