Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.53)  (0.47)
#> ===============================
#> V1
#>   mean           0.0382  0.0224
#>   std. dev.       0.029  0.0157
#>   weight sum         74      65
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2569  0.1794
#>   std. dev.      0.1494  0.1236
#>   weight sum         74      65
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean            0.296  0.1902
#>   std. dev.      0.1361  0.1252
#>   weight sum         74      65
#>   precision      0.0051  0.0051
#> 
#> V12
#>   mean           0.2991  0.2068
#>   std. dev.      0.1303   0.145
#>   weight sum         74      65
#>   precision       0.005   0.005
#> 
#> V13
#>   mean            0.312  0.2473
#>   std. dev.      0.1438   0.146
#>   weight sum         74      65
#>   precision      0.0052  0.0052
#> 
#> V14
#>   mean            0.329  0.2976
#>   std. dev.      0.1649  0.1781
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V15
#>   mean           0.3422  0.3449
#>   std. dev.      0.1754  0.2287
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V16
#>   mean           0.3857  0.4136
#>   std. dev.      0.2026  0.2665
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V17
#>   mean           0.4237  0.4463
#>   std. dev.        0.24  0.2955
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V18
#>   mean           0.4531  0.4785
#>   std. dev.      0.2527  0.2694
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V19
#>   mean           0.5398   0.494
#>   std. dev.      0.2464  0.2559
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V2
#>   mean           0.0501  0.0295
#>   std. dev.      0.0399  0.0206
#>   weight sum         74      65
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6198  0.5187
#>   std. dev.      0.2476   0.254
#>   weight sum         74      65
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean            0.674  0.5733
#>   std. dev.      0.2425  0.2495
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6604  0.6078
#>   std. dev.      0.2462  0.2447
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V23
#>   mean           0.6468  0.6415
#>   std. dev.      0.2669  0.2197
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6616  0.6762
#>   std. dev.      0.2602  0.2103
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V25
#>   mean           0.6573  0.6799
#>   std. dev.      0.2493  0.2187
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.6875  0.6866
#>   std. dev.      0.2357  0.2189
#>   weight sum         74      65
#>   precision      0.0065  0.0065
#> 
#> V27
#>   mean           0.7032  0.6746
#>   std. dev.      0.2574  0.2215
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V28
#>   mean           0.7031  0.6545
#>   std. dev.      0.2512  0.2264
#>   weight sum         74      65
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6496  0.6101
#>   std. dev.      0.2423  0.2585
#>   weight sum         74      65
#>   precision      0.0073  0.0073
#> 
#> V3
#>   mean           0.0555  0.0353
#>   std. dev.      0.0496  0.0274
#>   weight sum         74      65
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5841  0.5716
#>   std. dev.      0.2207  0.2492
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V31
#>   mean           0.4835  0.5336
#>   std. dev.      0.2182  0.1977
#>   weight sum         74      65
#>   precision      0.0063  0.0063
#> 
#> V32
#>   mean           0.4115  0.4653
#>   std. dev.      0.2031  0.2011
#>   weight sum         74      65
#>   precision      0.0063  0.0063
#> 
#> V33
#>   mean           0.3748  0.4404
#>   std. dev.      0.1709  0.2212
#>   weight sum         74      65
#>   precision      0.0068  0.0068
#> 
#> V34
#>   mean            0.358  0.4257
#>   std. dev.      0.1822  0.2473
#>   weight sum         74      65
#>   precision      0.0067  0.0067
#> 
#> V35
#>   mean           0.3361  0.4364
#>   std. dev.      0.2279  0.2556
#>   weight sum         74      65
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3079  0.4399
#>   std. dev.      0.2292  0.2477
#>   weight sum         74      65
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3146  0.3948
#>   std. dev.      0.2139  0.2285
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean           0.3303  0.3323
#>   std. dev.      0.1957  0.2152
#>   weight sum         74      65
#>   precision      0.0065  0.0065
#> 
#> V39
#>   mean           0.3348  0.3102
#>   std. dev.       0.182  0.2095
#>   weight sum         74      65
#>   precision       0.007   0.007
#> 
#> V4
#>   mean           0.0725  0.0431
#>   std. dev.      0.0618  0.0313
#>   weight sum         74      65
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3001  0.3276
#>   std. dev.      0.1496  0.1979
#>   weight sum         74      65
#>   precision      0.0066  0.0066
#> 
#> V41
#>   mean           0.2875  0.2874
#>   std. dev.      0.1611  0.1884
#>   weight sum         74      65
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean            0.311  0.2522
#>   std. dev.      0.1653  0.1722
#>   weight sum         74      65
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2896   0.212
#>   std. dev.      0.1354  0.1352
#>   weight sum         74      65
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2525  0.1864
#>   std. dev.      0.1396  0.1131
#>   weight sum         74      65
#>   precision      0.0057  0.0057
#> 
#> V45
#>   mean           0.2502  0.1515
#>   std. dev.       0.175   0.104
#>   weight sum         74      65
#>   precision      0.0051  0.0051
#> 
#> V46
#>   mean           0.2065  0.1222
#>   std. dev.      0.1571  0.1033
#>   weight sum         74      65
#>   precision      0.0053  0.0053
#> 
#> V47
#>   mean           0.1522  0.1002
#>   std. dev.      0.0991  0.0741
#>   weight sum         74      65
#>   precision       0.004   0.004
#> 
#> V48
#>   mean            0.116   0.071
#>   std. dev.       0.067  0.0517
#>   weight sum         74      65
#>   precision      0.0024  0.0024
#> 
#> V49
#>   mean           0.0665   0.039
#>   std. dev.      0.0361  0.0334
#>   weight sum         74      65
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0938  0.0633
#>   std. dev.      0.0678  0.0496
#>   weight sum         74      65
#>   precision      0.0029  0.0029
#> 
#> V50
#>   mean           0.0242  0.0187
#>   std. dev.      0.0145  0.0132
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0203  0.0134
#>   std. dev.      0.0147  0.0089
#>   weight sum         74      65
#>   precision      0.0008  0.0008
#> 
#> V52
#>   mean            0.017  0.0108
#>   std. dev.      0.0116   0.007
#>   weight sum         74      65
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0129  0.0091
#>   std. dev.      0.0083  0.0055
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean            0.013  0.0088
#>   std. dev.      0.0084  0.0049
#>   weight sum         74      65
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0103  0.0079
#>   std. dev.      0.0088  0.0046
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0094   0.008
#>   std. dev.      0.0062  0.0053
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0082  0.0084
#>   std. dev.      0.0063  0.0062
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0103  0.0065
#>   std. dev.      0.0083  0.0048
#>   weight sum         74      65
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0095  0.0069
#>   std. dev.      0.0074  0.0047
#>   weight sum         74      65
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1155  0.0983
#>   std. dev.      0.0571  0.0715
#>   weight sum         74      65
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean           0.0074  0.0061
#>   std. dev.      0.0065   0.004
#>   weight sum         74      65
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1279  0.1213
#>   std. dev.      0.0575  0.0717
#>   weight sum         74      65
#>   precision      0.0028  0.0028
#> 
#> V8
#>   mean           0.1464  0.1304
#>   std. dev.      0.0836  0.0885
#>   weight sum         74      65
#>   precision      0.0031  0.0031
#> 
#> V9
#>   mean            0.216  0.1498
#>   std. dev.      0.1233  0.1116
#>   weight sum         74      65
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3333333