Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$marshal()

LearnerClassifNaiveBayesWeka$unmarshal()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0346  0.0224
#>   std. dev.      0.0271  0.0131
#>   weight sum         73      66
#>   precision       0.001   0.001
#> 
#> V10
#>   mean            0.245  0.1594
#>   std. dev.       0.128  0.1054
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V11
#>   mean           0.2822  0.1802
#>   std. dev.      0.1142  0.1125
#>   weight sum         73      66
#>   precision      0.0047  0.0047
#> 
#> V12
#>   mean           0.2929  0.2041
#>   std. dev.      0.1222  0.1344
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3089  0.2407
#>   std. dev.      0.1208  0.1382
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V14
#>   mean           0.3216  0.2799
#>   std. dev.      0.1646  0.1714
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V15
#>   mean           0.3362  0.3091
#>   std. dev.      0.2063  0.2238
#>   weight sum         73      66
#>   precision      0.0074  0.0074
#> 
#> V16
#>   mean           0.3895  0.3819
#>   std. dev.      0.2219  0.2547
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V17
#>   mean           0.4216  0.4277
#>   std. dev.      0.2475   0.295
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4704  0.4659
#>   std. dev.      0.2542  0.2774
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean           0.5556  0.4828
#>   std. dev.       0.254  0.2647
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0464  0.0311
#>   std. dev.      0.0403  0.0253
#>   weight sum         73      66
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean            0.631  0.5101
#>   std. dev.      0.2566  0.2648
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6786  0.5446
#>   std. dev.      0.2504  0.2576
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V22
#>   mean           0.6804  0.5684
#>   std. dev.      0.2383   0.261
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6847  0.6163
#>   std. dev.      0.2496  0.2454
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V24
#>   mean           0.6826  0.6597
#>   std. dev.      0.2452   0.229
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6724  0.6705
#>   std. dev.      0.2462  0.2453
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.7029   0.688
#>   std. dev.       0.241  0.2397
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V27
#>   mean           0.7129  0.6719
#>   std. dev.      0.2703   0.218
#>   weight sum         73      66
#>   precision      0.0076  0.0076
#> 
#> V28
#>   mean            0.718  0.6472
#>   std. dev.      0.2611  0.2149
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V29
#>   mean            0.654  0.6177
#>   std. dev.      0.2401  0.2362
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V3
#>   mean           0.0522  0.0368
#>   std. dev.      0.0497  0.0289
#>   weight sum         73      66
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5779  0.5703
#>   std. dev.      0.1966  0.2325
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4633   0.528
#>   std. dev.       0.218   0.203
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4076  0.4559
#>   std. dev.      0.2114  0.2128
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V33
#>   mean           0.3784  0.4556
#>   std. dev.      0.2017  0.2196
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3511  0.4619
#>   std. dev.       0.214  0.2449
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V35
#>   mean           0.3307  0.4671
#>   std. dev.      0.2477  0.2523
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3095  0.4726
#>   std. dev.       0.259  0.2645
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3026  0.4302
#>   std. dev.      0.2347    0.25
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3267  0.3583
#>   std. dev.        0.21  0.2314
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V39
#>   mean           0.3425  0.3316
#>   std. dev.      0.1887  0.2254
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0705  0.0425
#>   std. dev.      0.0614  0.0277
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3007  0.3442
#>   std. dev.      0.1725   0.198
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2835  0.2997
#>   std. dev.      0.1632  0.1763
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.2898  0.2641
#>   std. dev.       0.167  0.1725
#>   weight sum         73      66
#>   precision      0.0057  0.0057
#> 
#> V43
#>   mean           0.2635  0.2221
#>   std. dev.      0.1436  0.1404
#>   weight sum         73      66
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2358  0.1839
#>   std. dev.      0.1377  0.1169
#>   weight sum         73      66
#>   precision      0.0058  0.0058
#> 
#> V45
#>   mean           0.2399  0.1491
#>   std. dev.      0.1674  0.1069
#>   weight sum         73      66
#>   precision      0.0051  0.0051
#> 
#> V46
#>   mean           0.2016  0.1205
#>   std. dev.      0.1474  0.0999
#>   weight sum         73      66
#>   precision      0.0053  0.0053
#> 
#> V47
#>   mean           0.1537  0.0988
#>   std. dev.      0.0946  0.0703
#>   weight sum         73      66
#>   precision      0.0039  0.0039
#> 
#> V48
#>   mean           0.1165  0.0724
#>   std. dev.      0.0685  0.0495
#>   weight sum         73      66
#>   precision      0.0024  0.0024
#> 
#> V49
#>   mean           0.0668  0.0401
#>   std. dev.      0.0368  0.0318
#>   weight sum         73      66
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0957  0.0646
#>   std. dev.      0.0659  0.0449
#>   weight sum         73      66
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0232  0.0187
#>   std. dev.      0.0146  0.0135
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0196  0.0129
#>   std. dev.      0.0131  0.0087
#>   weight sum         73      66
#>   precision      0.0008  0.0008
#> 
#> V52
#>   mean           0.0157  0.0104
#>   std. dev.      0.0103  0.0074
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0116  0.0091
#>   std. dev.      0.0078  0.0055
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0118  0.0092
#>   std. dev.      0.0083  0.0053
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean             0.01  0.0089
#>   std. dev.      0.0084  0.0054
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean            0.009  0.0077
#>   std. dev.       0.007   0.005
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0084  0.0076
#>   std. dev.      0.0061  0.0053
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0087  0.0067
#>   std. dev.      0.0072   0.005
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0086  0.0075
#>   std. dev.       0.007  0.0057
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V6
#>   mean           0.1138  0.1007
#>   std. dev.      0.0539  0.0673
#>   weight sum         73      66
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean           0.0072  0.0064
#>   std. dev.      0.0066  0.0039
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean            0.135  0.1203
#>   std. dev.      0.0614  0.0655
#>   weight sum         73      66
#>   precision      0.0026  0.0026
#> 
#> V8
#>   mean           0.1468   0.123
#>   std. dev.      0.0852   0.078
#>   weight sum         73      66
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2021  0.1348
#>   std. dev.      0.1177  0.1006
#>   weight sum         73      66
#>   precision      0.0048  0.0048
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551