Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.52)  (0.48)
#> ===============================
#> V1
#>   mean           0.0354   0.024
#>   std. dev.      0.0271  0.0163
#>   weight sum         73      66
#>   precision       0.001   0.001
#> 
#> V10
#>   mean            0.273  0.1643
#>   std. dev.      0.1427  0.1167
#>   weight sum         73      66
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean            0.306  0.1907
#>   std. dev.      0.1197  0.1221
#>   weight sum         73      66
#>   precision      0.0053  0.0053
#> 
#> V12
#>   mean           0.3048  0.2228
#>   std. dev.      0.1195   0.143
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V13
#>   mean           0.3147  0.2614
#>   std. dev.       0.123  0.1432
#>   weight sum         73      66
#>   precision       0.005   0.005
#> 
#> V14
#>   mean            0.321   0.291
#>   std. dev.      0.1691   0.172
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3306  0.3279
#>   std. dev.      0.1976  0.2256
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V16
#>   mean           0.3742   0.397
#>   std. dev.      0.2072  0.2553
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V17
#>   mean           0.4088  0.4309
#>   std. dev.      0.2287  0.2986
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V18
#>   mean           0.4524  0.4637
#>   std. dev.      0.2475  0.2851
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V19
#>   mean            0.529  0.4853
#>   std. dev.      0.2558  0.2771
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V2
#>   mean           0.0444  0.0302
#>   std. dev.      0.0383  0.0244
#>   weight sum         73      66
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6208  0.5207
#>   std. dev.      0.2577  0.2707
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6697  0.5423
#>   std. dev.      0.2569  0.2498
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V22
#>   mean            0.669  0.5385
#>   std. dev.      0.2395  0.2732
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V23
#>   mean           0.6755  0.5774
#>   std. dev.      0.2495  0.2428
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6857  0.6191
#>   std. dev.      0.2417   0.241
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6819  0.6393
#>   std. dev.      0.2437  0.2693
#>   weight sum         73      66
#>   precision      0.0073  0.0073
#> 
#> V26
#>   mean           0.7265  0.6648
#>   std. dev.      0.2324  0.2497
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V27
#>   mean           0.7342  0.6683
#>   std. dev.      0.2604  0.2149
#>   weight sum         73      66
#>   precision      0.0074  0.0074
#> 
#> V28
#>   mean           0.7297  0.6785
#>   std. dev.      0.2546  0.1928
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V29
#>   mean           0.6667  0.6552
#>   std. dev.      0.2385  0.2182
#>   weight sum         73      66
#>   precision      0.0075  0.0075
#> 
#> V3
#>   mean           0.0476  0.0356
#>   std. dev.      0.0426  0.0289
#>   weight sum         73      66
#>   precision      0.0023  0.0023
#> 
#> V30
#>   mean           0.5998  0.5864
#>   std. dev.      0.1949  0.2316
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V31
#>   mean           0.4932  0.5168
#>   std. dev.      0.2196   0.202
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4392  0.4318
#>   std. dev.      0.2103  0.2129
#>   weight sum         73      66
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4026  0.4289
#>   std. dev.       0.188  0.2191
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V34
#>   mean           0.3708  0.4324
#>   std. dev.      0.2129  0.2511
#>   weight sum         73      66
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3429  0.4258
#>   std. dev.      0.2581  0.2561
#>   weight sum         73      66
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3365  0.4276
#>   std. dev.      0.2621  0.2592
#>   weight sum         73      66
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean            0.331  0.3848
#>   std. dev.      0.2308  0.2429
#>   weight sum         73      66
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean           0.3489  0.3195
#>   std. dev.      0.2092  0.2174
#>   weight sum         73      66
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3572  0.3037
#>   std. dev.      0.1858  0.2103
#>   weight sum         73      66
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0642  0.0424
#>   std. dev.      0.0559  0.0309
#>   weight sum         73      66
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3124  0.3195
#>   std. dev.      0.1536  0.1953
#>   weight sum         73      66
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2881  0.2887
#>   std. dev.      0.1493  0.1746
#>   weight sum         73      66
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.3067  0.2512
#>   std. dev.      0.1674  0.1634
#>   weight sum         73      66
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean            0.284  0.2054
#>   std. dev.      0.1461  0.1229
#>   weight sum         73      66
#>   precision      0.0055  0.0055
#> 
#> V44
#>   mean           0.2544   0.163
#>   std. dev.      0.1431  0.0858
#>   weight sum         73      66
#>   precision      0.0043  0.0043
#> 
#> V45
#>   mean           0.2568  0.1337
#>   std. dev.      0.1762  0.0795
#>   weight sum         73      66
#>   precision      0.0051  0.0051
#> 
#> V46
#>   mean           0.1975  0.1142
#>   std. dev.       0.153  0.0917
#>   weight sum         73      66
#>   precision      0.0055  0.0055
#> 
#> V47
#>   mean           0.1398  0.0932
#>   std. dev.      0.0881  0.0677
#>   weight sum         73      66
#>   precision      0.0041  0.0041
#> 
#> V48
#>   mean           0.1113  0.0668
#>   std. dev.      0.0681  0.0478
#>   weight sum         73      66
#>   precision      0.0025  0.0025
#> 
#> V49
#>   mean           0.0643  0.0365
#>   std. dev.      0.0359   0.031
#>   weight sum         73      66
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0896  0.0618
#>   std. dev.      0.0606  0.0472
#>   weight sum         73      66
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0229  0.0182
#>   std. dev.       0.014  0.0136
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0188  0.0121
#>   std. dev.      0.0137  0.0086
#>   weight sum         73      66
#>   precision      0.0009  0.0009
#> 
#> V52
#>   mean           0.0153  0.0105
#>   std. dev.      0.0106  0.0074
#>   weight sum         73      66
#>   precision      0.0007  0.0007
#> 
#> V53
#>   mean           0.0115    0.01
#>   std. dev.      0.0077  0.0062
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0118    0.01
#>   std. dev.       0.008  0.0056
#>   weight sum         73      66
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0099  0.0093
#>   std. dev.      0.0083  0.0054
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V56
#>   mean           0.0089  0.0078
#>   std. dev.      0.0069  0.0047
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0075  0.0081
#>   std. dev.      0.0059   0.006
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0091  0.0071
#>   std. dev.      0.0076   0.005
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0085  0.0073
#>   std. dev.      0.0068  0.0053
#>   weight sum         73      66
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1184  0.0977
#>   std. dev.      0.0503  0.0651
#>   weight sum         73      66
#>   precision      0.0028  0.0028
#> 
#> V60
#>   mean            0.007   0.006
#>   std. dev.      0.0065  0.0034
#>   weight sum         73      66
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1372  0.1181
#>   std. dev.      0.0546  0.0677
#>   weight sum         73      66
#>   precision      0.0027  0.0027
#> 
#> V8
#>   mean           0.1595  0.1219
#>   std. dev.      0.0898  0.0877
#>   weight sum         73      66
#>   precision      0.0034  0.0034
#> 
#> V9
#>   mean           0.2314  0.1383
#>   std. dev.      0.1315  0.1046
#>   weight sum         73      66
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551