Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.naive_bayes_weka")
print(learner)
#> <LearnerClassifNaiveBayesWeka:classif.naive_bayes_weka>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings, multiclass, twoclass

# Define a Task
task = mlr3::tsk("sonar")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.55)  (0.45)
#> ===============================
#> V1
#>   mean            0.036  0.0236
#>   std. dev.      0.0267  0.0162
#>   weight sum         77      62
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2438   0.155
#>   std. dev.        0.14  0.0933
#>   weight sum         77      62
#>   precision      0.0047  0.0047
#> 
#> V11
#>   mean           0.2764  0.1718
#>   std. dev.       0.122  0.0947
#>   weight sum         77      62
#>   precision      0.0043  0.0043
#> 
#> V12
#>   mean           0.2895   0.184
#>   std. dev.      0.1227  0.1269
#>   weight sum         77      62
#>   precision      0.0046  0.0046
#> 
#> V13
#>   mean           0.3076   0.216
#>   std. dev.      0.1257  0.1375
#>   weight sum         77      62
#>   precision      0.0053  0.0053
#> 
#> V14
#>   mean           0.3161  0.2689
#>   std. dev.      0.1336  0.1674
#>   weight sum         77      62
#>   precision      0.0059  0.0059
#> 
#> V15
#>   mean            0.332  0.3053
#>   std. dev.      0.1733  0.2231
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V16
#>   mean           0.3927  0.3702
#>   std. dev.      0.1997  0.2571
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V17
#>   mean           0.4298  0.4222
#>   std. dev.      0.2336  0.2947
#>   weight sum         77      62
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4699  0.4586
#>   std. dev.      0.2601  0.2684
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V19
#>   mean           0.5548  0.4654
#>   std. dev.      0.2566  0.2581
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V2
#>   mean            0.045  0.0292
#>   std. dev.      0.0355  0.0237
#>   weight sum         77      62
#>   precision      0.0013  0.0013
#> 
#> V20
#>   mean           0.6446   0.494
#>   std. dev.      0.2509  0.2525
#>   weight sum         77      62
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6954  0.5331
#>   std. dev.      0.2313  0.2444
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V22
#>   mean           0.6866   0.577
#>   std. dev.      0.2246  0.2525
#>   weight sum         77      62
#>   precision      0.0067  0.0067
#> 
#> V23
#>   mean           0.6841  0.6369
#>   std. dev.      0.2599  0.2356
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V24
#>   mean           0.6974  0.6821
#>   std. dev.      0.2462  0.2252
#>   weight sum         77      62
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6878  0.7016
#>   std. dev.      0.2446  0.2246
#>   weight sum         77      62
#>   precision      0.0071  0.0071
#> 
#> V26
#>   mean           0.6995  0.7068
#>   std. dev.       0.242  0.2284
#>   weight sum         77      62
#>   precision      0.0071  0.0071
#> 
#> V27
#>   mean           0.6923  0.6889
#>   std. dev.      0.2795  0.2177
#>   weight sum         77      62
#>   precision      0.0073  0.0073
#> 
#> V28
#>   mean           0.6911  0.6748
#>   std. dev.      0.2625  0.2182
#>   weight sum         77      62
#>   precision      0.0074  0.0074
#> 
#> V29
#>   mean           0.6291   0.658
#>   std. dev.      0.2424  0.2251
#>   weight sum         77      62
#>   precision      0.0075  0.0075
#> 
#> V3
#>   mean           0.0519  0.0358
#>   std. dev.      0.0381  0.0289
#>   weight sum         77      62
#>   precision      0.0015  0.0015
#> 
#> V30
#>   mean           0.5747  0.5975
#>   std. dev.      0.2021   0.207
#>   weight sum         77      62
#>   precision      0.0064  0.0064
#> 
#> V31
#>   mean           0.4732  0.5241
#>   std. dev.      0.2098  0.1951
#>   weight sum         77      62
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean           0.4271  0.4383
#>   std. dev.      0.2045  0.2215
#>   weight sum         77      62
#>   precision      0.0065  0.0065
#> 
#> V33
#>   mean           0.4059  0.4357
#>   std. dev.      0.1975  0.2123
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3776  0.4276
#>   std. dev.      0.2147  0.2494
#>   weight sum         77      62
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3514  0.4368
#>   std. dev.       0.251  0.2484
#>   weight sum         77      62
#>   precision      0.0072  0.0072
#> 
#> V36
#>   mean           0.3301  0.4487
#>   std. dev.      0.2482  0.2578
#>   weight sum         77      62
#>   precision      0.0073  0.0073
#> 
#> V37
#>   mean           0.3257  0.4072
#>   std. dev.      0.2256  0.2422
#>   weight sum         77      62
#>   precision      0.0066  0.0066
#> 
#> V38
#>   mean            0.348  0.3282
#>   std. dev.      0.2096  0.2191
#>   weight sum         77      62
#>   precision       0.007   0.007
#> 
#> V39
#>   mean           0.3522  0.2929
#>   std. dev.      0.1868  0.2073
#>   weight sum         77      62
#>   precision      0.0069  0.0069
#> 
#> V4
#>   mean           0.0638  0.0399
#>   std. dev.      0.0429  0.0274
#>   weight sum         77      62
#>   precision       0.002   0.002
#> 
#> V40
#>   mean           0.3223  0.2873
#>   std. dev.      0.1576  0.1741
#>   weight sum         77      62
#>   precision      0.0065  0.0065
#> 
#> V41
#>   mean           0.3185  0.2615
#>   std. dev.      0.1601  0.1442
#>   weight sum         77      62
#>   precision      0.0052  0.0052
#> 
#> V42
#>   mean           0.3113   0.239
#>   std. dev.      0.1641  0.1363
#>   weight sum         77      62
#>   precision      0.0059  0.0059
#> 
#> V43
#>   mean           0.2685   0.205
#>   std. dev.      0.1321  0.1118
#>   weight sum         77      62
#>   precision      0.0046  0.0046
#> 
#> V44
#>   mean           0.2426  0.1687
#>   std. dev.      0.1326  0.0914
#>   weight sum         77      62
#>   precision      0.0043  0.0043
#> 
#> V45
#>   mean           0.2451  0.1336
#>   std. dev.      0.1749  0.0855
#>   weight sum         77      62
#>   precision      0.0052  0.0052
#> 
#> V46
#>   mean           0.1999  0.1133
#>   std. dev.      0.1478   0.076
#>   weight sum         77      62
#>   precision      0.0047  0.0047
#> 
#> V47
#>   mean           0.1433  0.0904
#>   std. dev.      0.0902  0.0548
#>   weight sum         77      62
#>   precision      0.0032  0.0032
#> 
#> V48
#>   mean           0.1068  0.0684
#>   std. dev.      0.0655   0.044
#>   weight sum         77      62
#>   precision      0.0022  0.0022
#> 
#> V49
#>   mean           0.0613  0.0381
#>   std. dev.      0.0365  0.0279
#>   weight sum         77      62
#>   precision      0.0014  0.0014
#> 
#> V5
#>   mean           0.0863  0.0496
#>   std. dev.       0.052  0.0346
#>   weight sum         77      62
#>   precision      0.0024  0.0024
#> 
#> V50
#>   mean            0.023  0.0175
#>   std. dev.      0.0149  0.0115
#>   weight sum         77      62
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0187   0.012
#>   std. dev.      0.0121  0.0072
#>   weight sum         77      62
#>   precision      0.0007  0.0007
#> 
#> V52
#>   mean           0.0157  0.0096
#>   std. dev.      0.0104   0.005
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0117  0.0091
#>   std. dev.       0.008  0.0062
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0129  0.0091
#>   std. dev.       0.009  0.0057
#>   weight sum         77      62
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean             0.01  0.0077
#>   std. dev.      0.0084   0.005
#>   weight sum         77      62
#>   precision      0.0005  0.0005
#> 
#> V56
#>   mean           0.0096  0.0069
#>   std. dev.      0.0067  0.0043
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0085  0.0081
#>   std. dev.      0.0062  0.0053
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0091  0.0068
#>   std. dev.      0.0067  0.0048
#>   weight sum         77      62
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0091  0.0066
#>   std. dev.      0.0073  0.0051
#>   weight sum         77      62
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1162  0.0814
#>   std. dev.      0.0502  0.0514
#>   weight sum         77      62
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0068  0.0058
#>   std. dev.      0.0062  0.0037
#>   weight sum         77      62
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1296  0.1079
#>   std. dev.      0.0582  0.0562
#>   weight sum         77      62
#>   precision      0.0025  0.0025
#> 
#> V8
#>   mean           0.1555  0.1164
#>   std. dev.      0.0943  0.0642
#>   weight sum         77      62
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2163   0.141
#>   std. dev.      0.1354  0.0827
#>   weight sum         77      62
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3768116