Classification Naive Bayes Learner From Weka

Naive Bayes Classifier Using Estimator Classes. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.naive_bayes_weka")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

K	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

O	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

References

John GH, Langley P (1995). “Estimating Continuous Distributions in Bayesian Classifiers.” In Eleventh Conference on Uncertainty in Artificial Intelligence, 338-345.

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayesWeka

Active bindings

marshaled

(logical(1))
Whether the learner has been marshaled.

Methods

Public methods

LearnerClassifNaiveBayesWeka$new()

LearnerClassifNaiveBayesWeka$marshal()

LearnerClassifNaiveBayesWeka$unmarshal()

LearnerClassifNaiveBayesWeka$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayesWeka$new()

Method marshal()

Marshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$marshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::marshal_model().

Method unmarshal()

Unmarshal the learner's model.

Usage

LearnerClassifNaiveBayesWeka$unmarshal(...)

Arguments

...

(any)
Additional arguments passed to mlr3::unmarshal_model().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayesWeka$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.naive_bayes_weka")
print(learner)
#> 
#> ── <LearnerClassifNaiveBayesWeka> (classif.naive_bayes_weka): Naive Bayes ──────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and RWeka
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: marshal, missings, multiclass, and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Naive Bayes Classifier
#> 
#>                  Class
#> Attribute            M       R
#>                 (0.57)  (0.43)
#> ===============================
#> V1
#>   mean           0.0349  0.0218
#>   std. dev.      0.0283  0.0134
#>   weight sum         80      59
#>   precision      0.0011  0.0011
#> 
#> V10
#>   mean           0.2556  0.1621
#>   std. dev.      0.1492  0.1065
#>   weight sum         80      59
#>   precision      0.0051  0.0051
#> 
#> V11
#>   mean           0.2947  0.1771
#>   std. dev.      0.1341  0.1022
#>   weight sum         80      59
#>   precision      0.0052  0.0052
#> 
#> V12
#>   mean           0.2987  0.1758
#>   std. dev.      0.1281  0.1098
#>   weight sum         80      59
#>   precision      0.0039  0.0039
#> 
#> V13
#>   mean           0.3173  0.2225
#>   std. dev.      0.1337  0.1058
#>   weight sum         80      59
#>   precision      0.0051  0.0051
#> 
#> V14
#>   mean           0.3281  0.2733
#>   std. dev.       0.162   0.143
#>   weight sum         80      59
#>   precision      0.0071  0.0071
#> 
#> V15
#>   mean           0.3411   0.319
#>   std. dev.       0.198  0.2044
#>   weight sum         80      59
#>   precision      0.0067  0.0067
#> 
#> V16
#>   mean           0.3958  0.4077
#>   std. dev.      0.2173  0.2503
#>   weight sum         80      59
#>   precision      0.0069  0.0069
#> 
#> V17
#>   mean            0.426  0.4595
#>   std. dev.      0.2447  0.2893
#>   weight sum         80      59
#>   precision      0.0071  0.0071
#> 
#> V18
#>   mean           0.4696  0.4972
#>   std. dev.      0.2544  0.2686
#>   weight sum         80      59
#>   precision      0.0068  0.0068
#> 
#> V19
#>   mean           0.5607  0.5267
#>   std. dev.      0.2418  0.2514
#>   weight sum         80      59
#>   precision      0.0068  0.0068
#> 
#> V2
#>   mean            0.046    0.03
#>   std. dev.      0.0405  0.0247
#>   weight sum         80      59
#>   precision      0.0018  0.0018
#> 
#> V20
#>   mean           0.6451  0.5505
#>   std. dev.      0.2378  0.2637
#>   weight sum         80      59
#>   precision      0.0069  0.0069
#> 
#> V21
#>   mean           0.6916  0.5922
#>   std. dev.      0.2367  0.2481
#>   weight sum         80      59
#>   precision      0.0071  0.0071
#> 
#> V22
#>   mean           0.6946  0.6183
#>   std. dev.      0.2378  0.2504
#>   weight sum         80      59
#>   precision      0.0066  0.0066
#> 
#> V23
#>   mean           0.6926  0.6381
#>   std. dev.       0.255  0.2428
#>   weight sum         80      59
#>   precision      0.0072  0.0072
#> 
#> V24
#>   mean           0.6946  0.6778
#>   std. dev.       0.239  0.2255
#>   weight sum         80      59
#>   precision      0.0073  0.0073
#> 
#> V25
#>   mean           0.6772   0.692
#>   std. dev.      0.2349  0.2285
#>   weight sum         80      59
#>   precision      0.0072  0.0072
#> 
#> V26
#>   mean           0.6959   0.711
#>   std. dev.      0.2381  0.2181
#>   weight sum         80      59
#>   precision      0.0069  0.0069
#> 
#> V27
#>   mean           0.7023  0.6796
#>   std. dev.      0.2679  0.2239
#>   weight sum         80      59
#>   precision      0.0076  0.0076
#> 
#> V28
#>   mean           0.7093  0.6339
#>   std. dev.      0.2573  0.1967
#>   weight sum         80      59
#>   precision      0.0071  0.0071
#> 
#> V29
#>   mean           0.6418  0.5886
#>   std. dev.      0.2404  0.2263
#>   weight sum         80      59
#>   precision      0.0074  0.0074
#> 
#> V3
#>   mean           0.0502  0.0356
#>   std. dev.      0.0454  0.0285
#>   weight sum         80      59
#>   precision      0.0024  0.0024
#> 
#> V30
#>   mean           0.5632  0.5428
#>   std. dev.       0.202  0.2278
#>   weight sum         80      59
#>   precision       0.007   0.007
#> 
#> V31
#>   mean           0.4568  0.4902
#>   std. dev.      0.2174  0.1912
#>   weight sum         80      59
#>   precision      0.0066  0.0066
#> 
#> V32
#>   mean             0.41  0.4215
#>   std. dev.      0.2042  0.2124
#>   weight sum         80      59
#>   precision      0.0066  0.0066
#> 
#> V33
#>   mean           0.3859  0.4245
#>   std. dev.      0.1886  0.1963
#>   weight sum         80      59
#>   precision       0.007   0.007
#> 
#> V34
#>   mean           0.3628  0.4243
#>   std. dev.      0.2097  0.2329
#>   weight sum         80      59
#>   precision      0.0068  0.0068
#> 
#> V35
#>   mean           0.3412  0.4361
#>   std. dev.       0.255  0.2451
#>   weight sum         80      59
#>   precision      0.0071  0.0071
#> 
#> V36
#>   mean           0.3299  0.4598
#>   std. dev.      0.2567  0.2509
#>   weight sum         80      59
#>   precision      0.0072  0.0072
#> 
#> V37
#>   mean           0.3306  0.4225
#>   std. dev.      0.2375  0.2426
#>   weight sum         80      59
#>   precision      0.0067  0.0067
#> 
#> V38
#>   mean           0.3501  0.3352
#>   std. dev.      0.2121  0.2173
#>   weight sum         80      59
#>   precision      0.0065  0.0065
#> 
#> V39
#>   mean           0.3631  0.2974
#>   std. dev.      0.1906  0.2079
#>   weight sum         80      59
#>   precision      0.0068  0.0068
#> 
#> V4
#>   mean           0.0652  0.0468
#>   std. dev.      0.0568  0.0326
#>   weight sum         80      59
#>   precision      0.0033  0.0033
#> 
#> V40
#>   mean           0.3248   0.314
#>   std. dev.      0.1648   0.186
#>   weight sum         80      59
#>   precision      0.0067  0.0067
#> 
#> V41
#>   mean           0.2851   0.284
#>   std. dev.      0.1608  0.1755
#>   weight sum         80      59
#>   precision      0.0063  0.0063
#> 
#> V42
#>   mean           0.2871  0.2483
#>   std. dev.      0.1552  0.1647
#>   weight sum         80      59
#>   precision      0.0058  0.0058
#> 
#> V43
#>   mean           0.2666  0.2035
#>   std. dev.      0.1261  0.1298
#>   weight sum         80      59
#>   precision      0.0056  0.0056
#> 
#> V44
#>   mean           0.2437  0.1741
#>   std. dev.      0.1395  0.1197
#>   weight sum         80      59
#>   precision      0.0058  0.0058
#> 
#> V45
#>   mean           0.2459  0.1476
#>   std. dev.      0.1765  0.1008
#>   weight sum         80      59
#>   precision      0.0051  0.0051
#> 
#> V46
#>   mean           0.1997  0.1313
#>   std. dev.      0.1448  0.1001
#>   weight sum         80      59
#>   precision      0.0046  0.0046
#> 
#> V47
#>   mean           0.1443   0.104
#>   std. dev.       0.086  0.0758
#>   weight sum         80      59
#>   precision      0.0032  0.0032
#> 
#> V48
#>   mean            0.108  0.0744
#>   std. dev.      0.0671  0.0545
#>   weight sum         80      59
#>   precision      0.0021  0.0021
#> 
#> V49
#>   mean           0.0614   0.043
#>   std. dev.      0.0358  0.0356
#>   weight sum         80      59
#>   precision      0.0015  0.0015
#> 
#> V5
#>   mean           0.0902  0.0712
#>   std. dev.      0.0612  0.0492
#>   weight sum         80      59
#>   precision       0.003   0.003
#> 
#> V50
#>   mean           0.0224    0.02
#>   std. dev.      0.0141  0.0145
#>   weight sum         80      59
#>   precision      0.0007  0.0007
#> 
#> V51
#>   mean           0.0182  0.0133
#>   std. dev.        0.01  0.0093
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V52
#>   mean           0.0156  0.0113
#>   std. dev.      0.0093  0.0079
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V53
#>   mean           0.0119  0.0098
#>   std. dev.      0.0075  0.0062
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V54
#>   mean           0.0123  0.0101
#>   std. dev.      0.0082  0.0057
#>   weight sum         80      59
#>   precision      0.0003  0.0003
#> 
#> V55
#>   mean           0.0104  0.0093
#>   std. dev.      0.0089  0.0053
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V56
#>   mean           0.0088  0.0073
#>   std. dev.      0.0068  0.0045
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V57
#>   mean           0.0078  0.0084
#>   std. dev.      0.0059  0.0058
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V58
#>   mean           0.0087  0.0071
#>   std. dev.      0.0068   0.005
#>   weight sum         80      59
#>   precision      0.0005  0.0005
#> 
#> V59
#>   mean           0.0089   0.007
#>   std. dev.      0.0071  0.0052
#>   weight sum         80      59
#>   precision      0.0004  0.0004
#> 
#> V6
#>   mean           0.1113  0.0989
#>   std. dev.      0.0569  0.0619
#>   weight sum         80      59
#>   precision      0.0022  0.0022
#> 
#> V60
#>   mean           0.0074  0.0058
#>   std. dev.      0.0064  0.0037
#>   weight sum         80      59
#>   precision      0.0005  0.0005
#> 
#> V7
#>   mean           0.1316  0.1161
#>   std. dev.      0.0622  0.0626
#>   weight sum         80      59
#>   precision      0.0024  0.0024
#> 
#> V8
#>   mean           0.1508  0.1169
#>   std. dev.      0.0926  0.0753
#>   weight sum         80      59
#>   precision      0.0033  0.0033
#> 
#> V9
#>   mean           0.2164  0.1367
#>   std. dev.      0.1299  0.0938
#>   weight sum         80      59
#>   precision      0.0049  0.0049
#> 
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551