Classification Random Forest Learner

Random forest for classification. Calls randomForest::randomForest() from randomForest.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.randomForest")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, randomForest

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

replace	logical	TRUE	TRUE, FALSE	-

classwt	untyped	NULL		-

cutoff	untyped	-		-

strata	untyped	-		-

sampsize	untyped	-		-

nodesize	integer	1		\([1, \infty)\)

maxnodes	integer	-		\([1, \infty)\)

importance	character	FALSE	accuracy, gini, none	-

localImp	logical	FALSE	TRUE, FALSE	-

proximity	logical	FALSE	TRUE, FALSE	-

oob.prox	logical	-	TRUE, FALSE	-

norm.votes	logical	TRUE	TRUE, FALSE	-

do.trace	logical	FALSE	TRUE, FALSE	-

keep.forest	logical	TRUE	TRUE, FALSE	-

keep.inbag	logical	FALSE	TRUE, FALSE	-

predict.all	logical	FALSE	TRUE, FALSE	-

nodes	logical	FALSE	TRUE, FALSE	-

References

Breiman, Leo (2001). “Random Forests.” Machine Learning, 45(1), 5–32. ISSN 1573-0565, doi:10.1023/A:1010933404324 .

See also

Author

pat-s

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifRandomForest

Methods

Public methods

LearnerClassifRandomForest$new()

LearnerClassifRandomForest$importance()

LearnerClassifRandomForest$oob_error()

LearnerClassifRandomForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifRandomForest$new()

Method importance()

The importance scores are extracted from the slot importance. Parameter 'importance' must be set to either "accuracy" or "gini".

Usage

LearnerClassifRandomForest$importance()

Returns

Named numeric().

Method oob_error()

OOB errors are extracted from the model slot err.rate.

Usage

LearnerClassifRandomForest$oob_error()

Returns

numeric(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifRandomForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.randomForest", importance = "accuracy")
print(learner)
#> 
#> ── <LearnerClassifRandomForest> (classif.randomForest): Random Forest ──────────
#> • Model: -
#> • Parameters: importance=accuracy
#> • Packages: mlr3, mlr3extralearners, and randomForest
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, multiclass, oob_error, twoclass, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")
# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> 
#> Call:
#>  randomForest(formula = formula, data = data, classwt = classwt,      cutoff = cutoff, importance = TRUE) 
#>                Type of random forest: classification
#>                      Number of trees: 500
#> No. of variables tried at each split: 7
#> 
#>         OOB estimate of  error rate: 23.74%
#> Confusion matrix:
#>    M  R class.error
#> M 59 14   0.1917808
#> R 19 47   0.2878788
print(learner$importance())
#>           V12           V11           V37           V51            V9 
#>  2.129655e-02  1.799237e-02  1.098511e-02  9.447100e-03  9.074140e-03 
#>           V36           V48           V49           V28           V13 
#>  8.975173e-03  7.374121e-03  7.146906e-03  6.221037e-03  6.016639e-03 
#>           V52           V16           V43           V10           V17 
#>  5.938297e-03  5.814378e-03  5.642745e-03  5.345010e-03  4.807645e-03 
#>           V47           V27            V4           V35           V59 
#>  4.621192e-03  4.442948e-03  4.440306e-03  3.900134e-03  3.653463e-03 
#>           V44           V15           V20           V54           V18 
#>  3.608148e-03  3.246376e-03  3.147665e-03  3.147128e-03  3.090163e-03 
#>           V45           V46           V21           V25           V26 
#>  2.713890e-03  2.487888e-03  2.458088e-03  2.194887e-03  2.077322e-03 
#>           V23           V55            V8            V3           V31 
#>  2.030867e-03  1.939048e-03  1.932185e-03  1.899767e-03  1.795170e-03 
#>           V42           V19           V22            V5           V29 
#>  1.744862e-03  1.585877e-03  1.557683e-03  1.516046e-03  1.478152e-03 
#>           V14           V39            V1           V41           V24 
#>  1.467654e-03  1.325776e-03  1.188350e-03  1.084973e-03  9.161486e-04 
#>           V30           V32           V58           V34           V53 
#>  8.038414e-04  7.425476e-04  7.288191e-04  5.606873e-04  5.377766e-04 
#>           V56           V38           V33           V50            V6 
#>  3.429099e-04  3.256889e-04  1.947808e-04  1.574474e-04  4.694906e-05 
#>           V60           V40            V7           V57            V2 
#> -7.816099e-05 -2.109994e-04 -2.711713e-04 -2.901363e-04 -6.377985e-04 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.115942