Classification Random Forest Learner

Random forest for classification. Calls randomForest::randomForest() from randomForest.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.randomForest")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3extralearners, randomForest

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

replace	logical	TRUE	TRUE, FALSE	-

classwt	untyped	NULL		-

cutoff	untyped	-		-

strata	untyped	-		-

sampsize	untyped	-		-

nodesize	integer	1		\([1, \infty)\)

maxnodes	integer	-		\([1, \infty)\)

importance	character	FALSE	accuracy, gini, none	-

localImp	logical	FALSE	TRUE, FALSE	-

proximity	logical	FALSE	TRUE, FALSE	-

oob.prox	logical	-	TRUE, FALSE	-

norm.votes	logical	TRUE	TRUE, FALSE	-

do.trace	logical	FALSE	TRUE, FALSE	-

keep.forest	logical	TRUE	TRUE, FALSE	-

keep.inbag	logical	FALSE	TRUE, FALSE	-

predict.all	logical	FALSE	TRUE, FALSE	-

nodes	logical	FALSE	TRUE, FALSE	-

References

Breiman, Leo (2001). “Random Forests.” Machine Learning, 45(1), 5–32. ISSN 1573-0565, doi:10.1023/A:1010933404324 .

See also

Author

pat-s

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifRandomForest

Methods

Public methods

LearnerClassifRandomForest$new()

LearnerClassifRandomForest$importance()

LearnerClassifRandomForest$oob_error()

LearnerClassifRandomForest$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifRandomForest$new()

Method importance()

The importance scores are extracted from the slot importance. Parameter 'importance' must be set to either "accuracy" or "gini".

Usage

LearnerClassifRandomForest$importance()

Returns

Named numeric().

Method oob_error()

OOB errors are extracted from the model slot err.rate.

Usage

LearnerClassifRandomForest$oob_error()

Returns

numeric(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifRandomForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("classif.randomForest", importance = "accuracy")
print(learner)
#> <LearnerClassifRandomForest:classif.randomForest>: Random Forest
#> * Model: -
#> * Parameters: importance=accuracy
#> * Packages: mlr3, mlr3extralearners, randomForest
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, multiclass, oob_error, twoclass, weights

# Define a Task
task = mlr3::tsk("sonar")
# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> 
#> Call:
#>  randomForest(formula = formula, data = data, classwt = classwt,      cutoff = cutoff, importance = TRUE) 
#>                Type of random forest: classification
#>                      Number of trees: 500
#> No. of variables tried at each split: 7
#> 
#>         OOB estimate of  error rate: 17.27%
#> Confusion matrix:
#>    M  R class.error
#> M 70  7  0.09090909
#> R 17 45  0.27419355
print(learner$importance())
#>            V9           V11           V48           V49           V12 
#>  2.222723e-02  1.604458e-02  1.506380e-02  1.447982e-02  1.295058e-02 
#>           V52           V47           V36           V10           V46 
#>  8.944927e-03  8.433634e-03  8.405009e-03  8.352065e-03  6.690651e-03 
#>           V21           V13           V37           V51           V22 
#>  5.801377e-03  4.883170e-03  4.646494e-03  4.529717e-03  4.378916e-03 
#>            V1           V44           V45            V5           V19 
#>  4.260711e-03  4.016248e-03  4.004358e-03  3.991649e-03  3.728054e-03 
#>           V28           V34           V31           V39           V23 
#>  3.278080e-03  3.029755e-03  2.787427e-03  2.702496e-03  2.577346e-03 
#>           V27           V33           V43           V35            V4 
#>  2.556624e-03  2.454094e-03  2.447300e-03  2.404445e-03  2.380972e-03 
#>           V20           V32           V30           V18           V15 
#>  1.932252e-03  1.892788e-03  1.822750e-03  1.600550e-03  1.397425e-03 
#>           V50           V55           V38           V24            V2 
#>  1.380544e-03  1.202641e-03  1.016606e-03  9.830515e-04  9.702407e-04 
#>           V17           V53            V7           V59           V16 
#>  9.562215e-04  9.474417e-04  8.557965e-04  8.125542e-04  7.751705e-04 
#>            V3           V14           V26           V40            V6 
#>  7.477845e-04  7.142403e-04  7.049202e-04  6.566195e-04  4.723225e-04 
#>           V58           V41           V29           V60           V54 
#>  3.696722e-04  3.297658e-04  2.721715e-04  1.519040e-04 -6.040534e-05 
#>           V25            V8           V42           V57           V56 
#> -1.715951e-04 -2.402349e-04 -3.962982e-04 -6.669670e-04 -8.580117e-04 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2173913