Classification Random Forest SRC Learner

Random forest for classification. Calls randomForestSRC::rfsrc() from randomForestSRC.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.rfsrc")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3extralearners, randomForestSRC

Parameters

Id	Type	Default	Levels	Range

ntree	integer	500		\([1, \infty)\)

mtry	integer	-		\([1, \infty)\)

mtry.ratio	numeric	-		\([0, 1]\)

nodesize	integer	15		\([1, \infty)\)

nodedepth	integer	-		\([1, \infty)\)

splitrule	character	gini	gini, auc, entropy	-

nsplit	integer	10		\([0, \infty)\)

importance	character	FALSE	FALSE, TRUE, none, permute, random, anti	-

block.size	integer	10		\([1, \infty)\)

bootstrap	character	by.root	by.root, by.node, none, by.user	-

samptype	character	swor	swor, swr	-

samp	untyped	-		-

membership	logical	FALSE	TRUE, FALSE	-

sampsize	untyped	-		-

sampsize.ratio	numeric	-		\([0, 1]\)

na.action	character	na.omit	na.omit, na.impute	-

nimpute	integer	1		\([1, \infty)\)

proximity	character	FALSE	FALSE, TRUE, inbag, oob, all	-

distance	character	FALSE	FALSE, TRUE, inbag, oob, all	-

forest.wt	character	FALSE	FALSE, TRUE, inbag, oob, all	-

xvar.wt	untyped	-		-

split.wt	untyped	-		-

forest	logical	TRUE	TRUE, FALSE	-

var.used	character	FALSE	FALSE, all.trees	-

split.depth	character	FALSE	FALSE, all.trees, by.tree	-

seed	integer	-		\((-\infty, -1]\)

do.trace	logical	FALSE	TRUE, FALSE	-

get.tree	untyped	-		-

outcome	character	train	train, test	-

ptn.count	integer	0		\([0, \infty)\)

cores	integer	1		\([1, \infty)\)

save.memory	logical	FALSE	TRUE, FALSE	-

perf.type	character	-	gmean, misclass, brier, none	-

case.depth	logical	FALSE	TRUE, FALSE	-

marginal.xvar	untyped	NULL		-

Custom mlr3 parameters

mtry: This hyperparameter can alternatively be set via the added hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

sampsize: This hyperparameter can alternatively be set via the added hyperparameter sampsize.ratio as sampsize = max(ceiling(sampsize.ratio * n_obs), 1). Note that sampsize and sampsize.ratio are mutually exclusive.

cores: This value is set as the option rf.cores during training and is set to 1 by default.

References

Breiman, Leo (2001). “Random Forests.” Machine Learning, 45(1), 5–32. ISSN 1573-0565, doi:10.1023/A:1010933404324 .

See also

Author

RaphaelS1

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifRandomForestSRC

Methods

Public methods

LearnerClassifRandomForestSRC$new()

LearnerClassifRandomForestSRC$importance()

LearnerClassifRandomForestSRC$selected_features()

LearnerClassifRandomForestSRC$oob_error()

LearnerClassifRandomForestSRC$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifRandomForestSRC$new()

Method importance()

The importance scores are extracted from the model slot importance, returned for 'all'.

Usage

LearnerClassifRandomForestSRC$importance()

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot var.used.

Note: Due to a known issue in randomForestSRC, enabling var.used = "all.trees" causes prediction to fail. Therefore, this setting should be used exclusively for feature selection purposes and not when prediction is required.

Usage

LearnerClassifRandomForestSRC$selected_features()

Returns

character().

Method oob_error()

OOB error extracted from the model slot err.rate.

Usage

LearnerClassifRandomForestSRC$oob_error()

Returns

numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifRandomForestSRC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.rfsrc", importance = "TRUE")
print(learner)
#> 
#> ── <LearnerClassifRandomForestSRC> (classif.rfsrc): Random Forest ──────────────
#> • Model: -
#> • Parameters: importance=TRUE
#> • Packages: mlr3, mlr3extralearners, and randomForestSRC
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, numeric, and factor
#> • Encapsulation: none (fallback: -)
#> • Properties: importance, missings, multiclass, oob_error, selected_features,
#> twoclass, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")
# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#>                          Sample size: 139
#>            Frequency of class labels: M=71, R=68
#>                      Number of trees: 500
#>            Forest terminal node size: 1
#>        Average no. of terminal nodes: 16.962
#> No. of variables tried at each split: 8
#>               Total no. of variables: 60
#>        Resampling used to grow trees: swor
#>     Resample size used to grow trees: 88
#>                             Analysis: RF-C
#>                               Family: class
#>                       Splitting rule: gini *random*
#>        Number of random split points: 10
#>                     Imbalanced ratio: 1.0441
#>                    (OOB) Brier score: 0.13352929
#>         (OOB) Normalized Brier score: 0.53411716
#>                            (OOB) AUC: 0.93371997
#>                       (OOB) Log-loss: 0.42891334
#>                         (OOB) PR-AUC: 0.92511334
#>                         (OOB) G-mean: 0.83670954
#>    (OOB) Requested performance error: 0.15827338, 0.08450704, 0.23529412
#> 
#> Confusion matrix:
#> 
#>           predicted
#>   observed  M  R class.error
#>          M 65  6      0.0845
#>          R 16 52      0.2353
#> 
#>       (OOB) Misclassification rate: 0.1582734
#> 
#> Random-classifier baselines (uniform):
#>    Brier: 0.25   Normalized Brier: 1   Log-loss: 0.69314718
print(learner$importance())
#>         V11         V12         V36         V52          V9         V37 
#> 0.099877018 0.083068833 0.035370801 0.033387049 0.031538775 0.025083242 
#>         V23         V18         V45         V10          V5         V17 
#> 0.024980005 0.022258314 0.021567770 0.020952051 0.020085590 0.019337999 
#>         V27          V7         V21         V51          V2          V4 
#> 0.016002044 0.015068514 0.014349605 0.012960216 0.012528717 0.012508921 
#>         V47         V48         V46         V39         V49         V13 
#> 0.011904413 0.010891542 0.010811510 0.010608743 0.010537409 0.010304084 
#>          V6         V16         V26         V24         V54         V44 
#> 0.010302170 0.010168971 0.009949718 0.009713948 0.009347114 0.009267438 
#>         V41         V29         V43         V20         V30         V35 
#> 0.009159632 0.008736500 0.008576623 0.008445405 0.008154477 0.007994657 
#>         V15         V22          V1         V34         V19         V31 
#> 0.007692060 0.007601422 0.007409799 0.006567708 0.006082426 0.005963689 
#>         V38         V55         V53         V50         V14         V56 
#> 0.005524404 0.005505343 0.005217777 0.005078744 0.005068600 0.004954200 
#>         V28          V3         V58          V8         V57         V42 
#> 0.004820194 0.004806363 0.004784892 0.004651804 0.004347761 0.004088960 
#>         V25         V33         V59         V40         V32         V60 
#> 0.004076301 0.003487699 0.003352782 0.002490465 0.002142868 0.001486852 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.1014493