Shrinkage Discriminant Analysis

Shrinkage Discriminant Analysis for classification. Calls sda::sda() from sda.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.sda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, sda

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	-		\([0, 1]\)

lambda.var	numeric	-		\([0, 1]\)

lambda.freqs	numeric	-		\([0, 1]\)

diagonal	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

References

Ahdesmaeki, Miika, Strimmer, Korbinian (2010). “Feature selection in omics prediction problems using cat scores and false nondiscovery rate control.” The Annals of Applied Statistics, 4(1). ISSN 1932-6157, doi:10.1214/09-aoas277 , http://dx.doi.org/10.1214/09-AOAS277.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSda

Methods

Public methods

LearnerClassifSda$new()

LearnerClassifSda$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSda$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSda$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.sda")
print(learner)
#> 
#> ── <LearnerClassifSda> (classif.sda): Shrinkage Discriminant Analysis ──────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and sda
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Number of variables: 60 
#> Number of observations: 139 
#> Number of classes: 2 
#> 
#> Estimating optimal shrinkage intensity lambda.freq (frequencies): 0.3102 
#> Estimating variances (pooled across classes)
#> Estimating optimal shrinkage intensity lambda.var (variance vector): 0.0255 
#> 
#> 
#> Computing inverse correlation matrix (pooled across classes)
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.1102 

print(learner$model)
#> $regularization
#>       lambda   lambda.var lambda.freqs 
#>   0.11023141   0.02546001   0.31023037 
#> 
#> $freqs
#>         M         R 
#> 0.5521049 0.4478951 
#> 
#> $alpha
#>         M         R 
#> -4.773069  2.930643 
#> 
#> $beta
#>           V1        V10       V11       V12        V13         V14        V15
#> M  0.7539186  0.2024422  1.594835  2.512688 -0.2300263  0.02659710  0.2275947
#> R -0.9293295 -0.2495435 -1.965899 -3.097304  0.2835456 -0.03278533 -0.2805482
#>            V16        V17         V18        V19        V2        V20
#> M -0.004630264 -0.3762731 -0.08514083  0.5890945  3.217556  0.7899038
#> R  0.005707568  0.4638189  0.10495018 -0.7261565 -3.966171 -0.9736873
#>          V21        V22        V23       V24        V25        V26        V27
#> M  0.2999555  0.1040227  0.7228811  1.143287 -0.8280931  0.0877929  0.1349948
#> R -0.3697448 -0.1282252 -0.8910706 -1.409291  1.0207618 -0.1082193 -0.1664035
#>          V28        V29        V3        V30       V31       V32        V33
#> M  0.1976028  0.5087789 -7.649323 -0.1971513 -2.475333  1.008581  0.8696485
#> R -0.2435781 -0.6271543  9.429057  0.2430216  3.051257 -1.243244 -1.0719858
#>          V34        V35        V36       V37       V38       V39        V4
#> M -0.5296893  0.4023650 -0.8837066 -1.831462  1.155044  1.759578  4.974621
#> R  0.6529297 -0.4959815  1.0893147  2.257580 -1.423783 -2.168972 -6.132044
#>         V40        V41        V42        V43         V44        V45       V46
#> M -1.866728  0.2774874  0.5347628  0.3785952  0.07499320  0.2490288  2.810369
#> R  2.301051 -0.3420492 -0.6591837 -0.4666812 -0.09244154 -0.3069692 -3.464245
#>         V47       V48       V49        V5       V50       V51       V52
#> M  4.366844  3.049651   9.14448  1.600620 -17.25435 -7.031258 -1.170159
#> R -5.382858 -3.759200 -11.27209 -1.973029  21.26885  8.667190  1.442415
#>         V53       V54      V55       V56       V57        V58        V59
#> M  2.020667  2.318092 -9.38246  2.345621 -4.308006  0.9572406  0.2383694
#> R -2.490807 -2.857432 11.56544 -2.891367  5.310331 -1.1799576 -0.2938298
#>            V6       V60        V7        V8        V9
#> M -0.01676818 -2.258862 -2.561612 -1.915265  2.891766
#> R  0.02066956  2.784422  3.157611  2.360881 -3.564580
#> attr(,"class")
#> [1] "shrinkage"
#> 
#> attr(,"class")
#> [1] "sda"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Prediction uses 60 features.

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2608696