Shrinkage Discriminant Analysis

Shrinkage Discriminant Analysis for classification. Calls sda::sda() from sda.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.sda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, sda

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	-		\([0, 1]\)

lambda.var	numeric	-		\([0, 1]\)

lambda.freqs	numeric	-		\([0, 1]\)

diagonal	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

References

Ahdesmaeki, Miika, Strimmer, Korbinian (2010). “Feature selection in omics prediction problems using cat scores and false nondiscovery rate control.” The Annals of Applied Statistics, 4(1). ISSN 1932-6157, doi:10.1214/09-aoas277 , http://dx.doi.org/10.1214/09-AOAS277.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSda

Methods

Public methods

LearnerClassifSda$new()

LearnerClassifSda$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSda$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSda$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.sda")
print(learner)
#> 
#> ── <LearnerClassifSda> (classif.sda): Shrinkage Discriminant Analysis ──────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and sda
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Number of variables: 60 
#> Number of observations: 139 
#> Number of classes: 2 
#> 
#> Estimating optimal shrinkage intensity lambda.freq (frequencies): 0.3806 
#> Estimating variances (pooled across classes)
#> Estimating optimal shrinkage intensity lambda.var (variance vector): 0.0214 
#> 
#> 
#> Computing inverse correlation matrix (pooled across classes)
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.1156 

print(learner$model)
#> $regularization
#>       lambda   lambda.var lambda.freqs 
#>   0.11562337   0.02138619   0.38058533 
#> 
#> $freqs
#>         M         R 
#> 0.5423341 0.4576659 
#> 
#> $alpha
#>          M          R 
#> -2.5745143  0.3642591 
#> 
#> $beta
#>            V1         V10       V11       V12        V13        V14        V15
#> M  0.02484557 -0.08093926  2.172570  1.603861  0.2290700  0.6139616 -0.5361372
#> R -0.02944200  0.09591303 -2.574496 -1.900575 -0.2714479 -0.7275445  0.6353225
#>          V16        V17        V18         V19        V2        V20        V21
#> M -0.8534199 -0.7881935  0.1999401  0.01295299  5.600379  0.4292988  0.2401972
#> R  1.0113026  0.9340093 -0.2369290 -0.01534929 -6.636450 -0.5087191 -0.2846336
#>          V22        V23        V24        V25        V26        V27        V28
#> M  0.3480735  0.7867862  0.6406404 -0.7847248 -0.1967538  0.3860286 -0.5109312
#> R -0.4124671 -0.9323416 -0.7591589  0.9298989  0.2331533 -0.4574439  0.6054535
#>          V29        V3        V30       V31       V32        V33        V34
#> M  0.8516656 -5.864957  0.6436128 -2.997895  1.069239  0.2415835 -0.7620415
#> R -1.0092237  6.949974 -0.7626812  3.552506 -1.267048 -0.2862764  0.9030192
#>          V35       V36       V37       V38        V39        V4       V40
#> M  0.1748584 -1.325244 -1.349846  1.210183  0.6149263  4.414447 -1.790412
#> R -0.2072073  1.570414  1.599568 -1.434067 -0.7286876 -5.231120  2.121638
#>          V41         V42       V43        V44       V45        V46         V47
#> M  0.9535081 -0.09305658  1.034519  0.9016337  1.616905  0.4423631 -0.01263383
#> R -1.1299071  0.11027205 -1.225906 -1.0684359 -1.916033 -0.5242003  0.01497109
#>         V48       V49        V5       V50       V51       V52       V53
#> M  2.871401  6.566043  1.240537 -6.940508 -3.151115  2.467596 -2.569203
#> R -3.402610 -7.780761 -1.470036  8.224502  3.734071 -2.924102  3.044505
#>         V54       V55        V56       V57       V58       V59         V6
#> M  4.321698 -11.12294 -0.5489442 -2.802493  5.255383  6.976295  0.3374028
#> R -5.121212  13.18069  0.6504989  3.320954 -6.227629 -8.266910 -0.3998223
#>         V60        V7        V8        V9
#> M -4.385621 -3.235326 -2.675114  2.521328
#> R  5.196960  3.833862  3.170010 -2.987773
#> attr(,"class")
#> [1] "shrinkage"
#> 
#> attr(,"class")
#> [1] "sda"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Prediction uses 60 features.

# Score the predictions
predictions$score()
#> classif.ce 
#>   0.173913