Shrinkage Discriminant Analysis

Shrinkage Discriminant Analysis for classification. Calls sda::sda() from sda.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.sda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, sda

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	-		\([0, 1]\)

lambda.var	numeric	-		\([0, 1]\)

lambda.freqs	numeric	-		\([0, 1]\)

diagonal	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

References

Ahdesmaeki, Miika, Strimmer, Korbinian (2010). “Feature selection in omics prediction problems using cat scores and false nondiscovery rate control.” The Annals of Applied Statistics, 4(1). ISSN 1932-6157, doi:10.1214/09-aoas277 , http://dx.doi.org/10.1214/09-AOAS277.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSda

Methods

Public methods

LearnerClassifSda$new()

LearnerClassifSda$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSda$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSda$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.sda")
print(learner)
#> 
#> ── <LearnerClassifSda> (classif.sda): Shrinkage Discriminant Analysis ──────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and sda
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Number of variables: 60 
#> Number of observations: 139 
#> Number of classes: 2 
#> 
#> Estimating optimal shrinkage intensity lambda.freq (frequencies): 0.3806 
#> Estimating variances (pooled across classes)
#> Estimating optimal shrinkage intensity lambda.var (variance vector): 0.0233 
#> 
#> 
#> Computing inverse correlation matrix (pooled across classes)
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.1146 

print(learner$model)
#> $regularization
#>       lambda   lambda.var lambda.freqs 
#>   0.11456091   0.02325321   0.38058533 
#> 
#> $freqs
#>         M         R 
#> 0.5423341 0.4576659 
#> 
#> $alpha
#>         M         R 
#> -4.907587  3.090070 
#> 
#> $beta
#>          V1        V10        V11       V12       V13        V14        V15
#> M  7.598162  0.8344958  0.8091931  1.961048  1.292628 -0.1378841  0.5250107
#> R -9.003822 -0.9888775 -0.9588938 -2.323842 -1.531764  0.1633926 -0.6221377
#>          V16        V17        V18        V19        V2        V20        V21
#> M -0.6805159 -0.7565277 -0.1971041  0.5074925 -2.594445  0.8470032  0.1216828
#> R  0.8064114  0.8964854  0.2335684 -0.6013786  3.074417 -1.0036988 -0.1441941
#>          V22        V23       V24         V25        V26        V27        V28
#> M  0.4649895  0.8327013  1.013334 -0.02551776 -0.4165627  0.4043263  0.3717086
#> R -0.5510126 -0.9867511 -1.200800  0.03023854  0.4936268 -0.4791267 -0.4404747
#>          V29        V3        V30       V31        V32        V33        V34
#> M  0.2299671 -10.25357  0.6243480 -2.018830  0.4679562 -0.6167010 -0.4848781
#> R -0.2725110  12.15049 -0.7398524  2.392313 -0.5545281  0.7307907  0.5745805
#>         V35        V36        V37        V38        V39         V4       V40
#> M  1.326789 -0.8557694 -0.9152743  0.9057898  0.2978655   8.997394 -1.448761
#> R -1.572245  1.0140868  1.0846000 -1.0733609 -0.3529707 -10.661912  1.716782
#>          V41        V42       V43        V44        V45       V46       V47
#> M  0.4741522  0.3968981  1.227128 -0.6530414  0.8339646  1.653897  2.581798
#> R -0.5618704 -0.4703242 -1.454147  0.7738541 -0.9882480 -1.959868 -3.059430
#>         V48       V49         V5      V50       V51       V52       V53
#> M  5.992788  7.233450  0.4571564 -12.2638 -10.63159  1.806883 -3.273823
#> R -7.101454 -8.571639 -0.5417303  14.5326  12.59843 -2.141156  3.879480
#>         V54       V55        V56       V57       V58       V59        V6
#> M  5.056842 -5.438488  0.8095316 -8.957747  1.589645  5.535667 -1.849919
#> R -5.992357  6.444608 -0.9592949 10.614930 -1.883730 -6.559765  2.192154
#>          V60         V7        V8        V9
#> M  0.5551683 -0.4555698 -2.137052  1.558003
#> R -0.6578744  0.5398502  2.532407 -1.846234
#> attr(,"class")
#> [1] "shrinkage"
#> 
#> attr(,"class")
#> [1] "sda"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Prediction uses 60 features.

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551