Shrinkage Discriminant Analysis

Shrinkage Discriminant Analysis for classification. Calls sda::sda() from sda.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.sda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, sda

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	-		\([0, 1]\)

lambda.var	numeric	-		\([0, 1]\)

lambda.freqs	numeric	-		\([0, 1]\)

diagonal	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

References

Ahdesmaeki, Miika, Strimmer, Korbinian (2010). “Feature selection in omics prediction problems using cat scores and false nondiscovery rate control.” The Annals of Applied Statistics, 4(1). ISSN 1932-6157, doi:10.1214/09-aoas277 , http://dx.doi.org/10.1214/09-AOAS277.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSda

Methods

Public methods

LearnerClassifSda$new()

LearnerClassifSda$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSda$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSda$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.sda")
print(learner)
#> 
#> ── <LearnerClassifSda> (classif.sda): Shrinkage Discriminant Analysis ──────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and sda
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Number of variables: 60 
#> Number of observations: 139 
#> Number of classes: 2 
#> 
#> Estimating optimal shrinkage intensity lambda.freq (frequencies): 0.615 
#> Estimating variances (pooled across classes)
#> Estimating optimal shrinkage intensity lambda.var (variance vector): 0.0225 
#> 
#> 
#> Computing inverse correlation matrix (pooled across classes)
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.1046 

print(learner$model)
#> $regularization
#>       lambda   lambda.var lambda.freqs 
#>    0.1046063    0.0224790    0.6150081 
#> 
#> $freqs
#>         M         R 
#> 0.5207729 0.4792271 
#> 
#> $alpha
#>         M         R 
#> -5.797335  3.292535 
#> 
#> $beta
#>            V1        V10       V11       V12        V13        V14        V15
#> M -0.06507867 -0.8948335  3.024756  2.693402  0.6984555 -0.1956229  0.9338334
#> R  0.07072057  0.9724098 -3.286983 -2.926903 -0.7590071  0.2125821 -1.0147908
#>          V16        V17         V18        V19        V2        V20        V21
#> M -0.2834405 -0.6904795 -0.09098890  0.4608761  10.49243  0.6767179  0.2794711
#> R  0.3080130  0.7503397  0.09887705 -0.5008310 -11.40205 -0.7353849 -0.3036995
#>          V22       V23        V24       V25        V26        V27        V28
#> M  0.1550761  1.156940  0.5646969 -1.139458  0.8195132  0.5832615 -0.6552265
#> R -0.1685202 -1.257239 -0.6136525  1.238241 -0.8905597 -0.6338265  0.7120304
#>          V29        V3       V30       V31        V32        V33         V34
#> M  0.2055546 -8.864713  1.725276 -1.502744 -0.1590472  0.5934585 -0.07221644
#> R -0.2233748  9.633226 -1.874846  1.633022  0.1728356 -0.6449075  0.07847714
#>          V35       V36       V37        V38       V39        V4       V40
#> M  0.4954870 -1.553152 -1.671668  0.8560853  2.260086  7.848684 -2.258363
#> R -0.5384426  1.687800  1.816591 -0.9303024 -2.456021 -8.529115  2.454148
#>          V41        V42        V43       V44       V45       V46       V47
#> M  0.4269949 -0.6807453 -0.3160379  1.058124  1.502472  3.354502  2.947006
#> R -0.4640126  0.7397615  0.3434364 -1.149856 -1.632726 -3.645316 -3.202492
#>         V48        V49        V5       V50       V51       V52       V53
#> M  1.558644   9.398749  2.120558 -18.25215 -1.465105 -6.274749 -4.588851
#> R -1.693768 -10.213560 -2.304397  19.83450  1.592120  6.818729  4.986675
#>          V54       V55       V56       V57       V58       V59         V6
#> M  0.4146684 -14.31229  3.953368 -6.049944  4.171031  1.142149 -0.3529018
#> R -0.4506175  15.55308 -4.296099  6.574435 -4.532632 -1.241166  0.3834961
#>          V60        V7        V8        V9
#> M  0.7050940 -3.359326 -1.749833  2.678658
#> R -0.7662211  3.650558  1.901532 -2.910880
#> attr(,"class")
#> [1] "shrinkage"
#> 
#> attr(,"class")
#> [1] "sda"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Prediction uses 60 features.

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2898551