Shrinkage Discriminant Analysis

Shrinkage Discriminant Analysis for classification. Calls sda::sda() from sda.

Dictionary

This Learner can be instantiated via lrn():

lrn("classif.sda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, sda

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	-		\([0, 1]\)

lambda.var	numeric	-		\([0, 1]\)

lambda.freqs	numeric	-		\([0, 1]\)

diagonal	logical	FALSE	TRUE, FALSE	-

verbose	logical	FALSE	TRUE, FALSE	-

References

Ahdesmaeki, Miika, Strimmer, Korbinian (2010). “Feature selection in omics prediction problems using cat scores and false nondiscovery rate control.” The Annals of Applied Statistics, 4(1). ISSN 1932-6157, doi:10.1214/09-aoas277 , http://dx.doi.org/10.1214/09-AOAS277.

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifSda

Methods

Public methods

LearnerClassifSda$new()

LearnerClassifSda$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifSda$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifSda$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("classif.sda")
print(learner)
#> 
#> ── <LearnerClassifSda> (classif.sda): Shrinkage Discriminant Analysis ──────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and sda
#> • Predict Types: [response] and prob
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass and twoclass
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Number of variables: 60 
#> Number of observations: 139 
#> Number of classes: 2 
#> 
#> Estimating optimal shrinkage intensity lambda.freq (frequencies): 0.8212 
#> Estimating variances (pooled across classes)
#> Estimating optimal shrinkage intensity lambda.var (variance vector): 0.0229 
#> 
#> 
#> Computing inverse correlation matrix (pooled across classes)
#> Estimating optimal shrinkage intensity lambda (correlation matrix): 0.113 

print(learner$model)
#> $regularization
#>       lambda   lambda.var lambda.freqs 
#>   0.11302999   0.02285135   0.82119887 
#> 
#> $freqs
#>         M         R 
#> 0.5083612 0.4916388 
#> 
#> $alpha
#>         M         R 
#> -4.466708  2.019216 
#> 
#> $beta
#>          V1        V10       V11       V12        V13        V14        V15
#> M  1.548895 -0.3139227  2.427404  1.891936  0.1841104 -0.7933878 -0.6262074
#> R -1.601578  0.3246004 -2.509969 -1.956288 -0.1903727  0.8203737  0.6475069
#>          V16        V17        V18        V19        V2        V20        V21
#> M -0.4225409 -0.6130627  0.6571868  0.9347317  5.914303  0.3380626 -0.2130138
#> R  0.4369130  0.6339152 -0.6795401 -0.9665253 -6.115470 -0.3495613  0.2202592
#>          V22        V23        V24        V25        V26        V27         V28
#> M  0.2203312  0.6163575  0.5305109 -0.4068988  0.1132917 -0.1962279  0.07855095
#> R -0.2278254 -0.6373221 -0.5485555  0.4207389 -0.1171451  0.2029024 -0.08122275
#>          V29        V3       V30       V31         V32        V33         V34
#> M  0.2403771 -5.392062  1.540462 -1.523872 -0.01423310 -0.8558978 -0.06098152
#> R -0.2485532  5.575466 -1.592859  1.575704  0.01471722  0.8850100  0.06305572
#>          V35       V36       V37        V38       V39        V4        V40
#> M  0.8738968 -1.667657 -1.987157  0.4506950  1.251618  7.549870 -0.5391995
#> R -0.9036211  1.724380  2.054748 -0.4660248 -1.294190 -7.806668  0.5575397
#>          V41        V42       V43       V44        V45       V46       V47
#> M  0.5395996  0.1221650  2.199691  2.487174  0.4827566  1.246143  4.024977
#> R -0.5579533 -0.1263203 -2.274510 -2.571772 -0.4991769 -1.288528 -4.161881
#>         V48       V49         V5       V50       V51       V52       V53
#> M  3.307629  5.402086 -0.5643829 -17.19869 -9.210178 -4.284399  6.446945
#> R -3.420133 -5.585831  0.5835796  17.78368  9.523449  4.430127 -6.666229
#>         V54       V55     V56       V57       V58        V59         V6
#> M  4.828782 -8.360652 -5.1362 -12.18610  10.81498  0.5304045 -0.9661523
#> R -4.993026  8.645028  5.3109  12.60059 -11.18284 -0.5484454  0.9990146
#>         V60         V7        V8        V9
#> M  2.911782 -0.9038912 -1.099493  1.363240
#> R -3.010823  0.9346358  1.136891 -1.409609
#> attr(,"class")
#> [1] "shrinkage"
#> 
#> attr(,"class")
#> [1] "sda"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Prediction uses 60 features.

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.2318841