Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 14.92295                                                 fv = 23.34698                                                 fv = 17.34158                                                 fv = 171.1118                                                 fv = 31688.51                                                 fv = 10.54846                                                 fv = 590.0756                                                 fv = 1.340781e+154                                                 fv = 15.31951                                                 fv = 16.93653                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.48265                                                 fv = 13.22621                                                 fv = 8.666841                                                 fv = 717.408                                                 fv = 21046.2                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.98265 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5888.191                                                 fv = 1.340781e+154                                                 fv = 29.96399                                                 fv = 9.172103                                                 fv = 6.88387                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 891.5387                                                 fv = 11.80595                                                 fv = 10.92758                                                 fv = 12.401                                                 fv = 1185.39                                                 fv = 6.88387                                                 fv = 83.74826                                                 fv = 1.340781e+154                                                 fv = 8.792928                                                 fv = 7.111162                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 195.523                                                 fv = 494.8177                                                 fv = 857.9795                                                 fv = 9.617199                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.09407                                                 fv = 233.5569                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 88.4663                                                 fv = 12.11061                                                 fv = 4615.799                                                 fv = 6.88387                                                 fv = 6.88387                                                 fv = 6.88387                                                 fv = 6.88387                                                 fv = 8.403567                                                 fv = 1.340781e+154                                                 fv = 5031143                                                 fv = 6.937939                                                 fv = 1.340781e+154                                                 fv = 43.19441                                                 fv = 99659.77                                                 fv = 22.78525                                                 fv = 6.577107                                                 fv = 18.68605                                                 fv = 1.340781e+154                                                 fv = 9.691458                                                 fv = 321.968                                                 fv = 32.15237                                                 fv = 9.277877                                                 fv = 8.275715                                                 fv = 8.172977                                                 fv = 10.8521                                                 fv = 10.8521                                                 fv = 10.8521                                                 fv = 1.340781e+154                                                 fv = 7.211376                                                 fv = 6.378209                                                 fv = 115.8642                                                 fv = 1.340781e+154                                                 fv = 8.649035                                                 fv = 25.89705                                                 fv = 1.340781e+154                                                 fv = 10.51085                                                 fv = 11.73249                                                 fv = 9.267842                                                 fv = 6.378209                                                 fv = 6.084622                                                 fv = 6.084622                                                 fv = 1.340781e+154                                                 fv = 43.53716                                                 fv = 7.795832                                                 fv = 33.50743                                                 fv = 6.084622                                                 fv = 6.351992                                                 fv = 11.94517                                                 fv = 7.161272                                                 fv = 12.88575                                                 fv = 1.340781e+154                                                 fv = 6.084622                                                 fv = 6.084622                                                 fv = 6.084622                                                 fv = 6.084622                                                 fv = 7.165352                                                 fv = 9.563694                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 749093                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.288487                                                 fv = 12.07966                                                 fv = 34.97228                                                 fv = 9.664068                                                 fv = 321.486                                                 fv = 6.084622                                                 fv = 6.577107                                                 fv = 1.340781e+154                                                 fv = 8.210552                                                 fv = 6.084622                                                 fv = 7.054736                                                 fv = 5.987282                                                 fv = 1225.064                                                 fv = 1.340781e+154                                                 fv = 8.304357                                                 fv = 9.194292                                                 fv = 15.78534                                                 fv = 1.340781e+154                                                 fv = 7.748533                                                 fv = 5.062201                                                 fv = 63.45395                                                 fv = 1.340781e+154                                                 fv = 8.646588                                                 fv = 8.531497                                                 fv = 6.423782                                                 fv = 4.770471                                                 fv = 1.340781e+154                                                 fv = 11.70263                                                 fv = 10.97708                                                 fv = 21.45978                                                 fv = 21.45978                                                 fv = 11.87616                                                 fv = 22.81397                                                 fv = 4.770471                                                 fv = 80.07111                                                 fv = 4.770471                                                 fv = 13.20512                                                 fv = 1.340781e+154                                                 fv = 7.162675                                                 fv = 17.31909                                                 fv = 5.129618                                                 fv = 11.1655                                                 fv = 4.5838                                                 fv = 21.72211                                                 fv = 1.340781e+154                                                 fv = 9.72357                                                 fv = 5.006999                                                 fv = 6.938406                                                 fv = 6.571561                                                 fv = 5.271317                                                 fv = 4.730562                                                 fv = 9.280125                                                 fv = 5.191337                                                 fv = 3.703069                                                 fv = 22406.37                                                 fv = 1.340781e+154                                                 fv = 8.753114                                                 fv = 5640.991                                                 fv = 1565.272                                                 fv = 47.35788                                                 fv = 20.02498                                                 fv = 7.723431                                                 fv = 7.512172                                                 fv = 5.917701                                                 fv = 6.78294                                                 fv = 5.917701                                                 fv = 90.94065                                                 fv = 5.88237                                                 fv = 3.703069                                                 fv = 1.340781e+154                                                 fv = 14.28169                                                 fv = 27.09017                                                 fv = 10.11457                                                 fv = 8.14628                                                 fv = 1.340781e+154                                                 fv = 4.978661                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.81637                                                 fv = 10.70948                                                 fv = 9.054923                                                 fv = 1.340781e+154                                                 fv = 10.29428                                                 fv = 13.26019                                                 fv = 8.512881                                                 fv = 10.34154                                                 fv = 16.20477                                                 fv = 5.021425                                                 fv = 5.309893                                                 fv = 15.51568                                                 fv = 10.79013                                                 fv = 4.106122                                                 fv = 28.83299                                                 fv = 13.2978                                                 fv = 11.80908                                                 fv = 5.267626                                                 fv = 3.703069                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.876283                                                 fv = 7.198942                                                 fv = 4.941109                                                 fv = 3.703069                                                 fv = 5.127719                                                 fv = 7.880326                                                 fv = 3.703069                                                 fv = 5.196522                                                 fv = 7.812106                                                 fv = 5.669601                                                 fv = 7.880326                                                 fv = 9.326922                                                 fv = 9.610585                                                 fv = 1.340781e+154                                                 fv = 3.703069                                                 fv = 10.62373 run # 0: f=3.703069464
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.792454                                                 fv = 6.145991                                                 fv = 36.75275                                                 fv = 7.01629                                                 fv = 12.09328                                                 fv = 4.597743                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.800002                                                 fv = 6.190697                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2941120062                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.597743                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.50439                                                 fv = 20.8701                                                 fv = 10.81858                                                 fv = 5.16446                                                 fv = 14.53346                                                 fv = 1.340781e+154                                                 fv = 40268315                                                 fv = 6.507141                                                 fv = 13.94672                                                 fv = 8.303992                                                 fv = 23.07982                                                 fv = 4.616308                                                 fv = 4.597743                                                 fv = 1.340781e+154                                                 fv = 151576.8                                                 fv = 13.14876                                                 fv = 4.597743                                                 fv = 4.597743                                                 fv = 4.616308                                                 fv = 1.340781e+154                                                 fv = 17.54575                                                 fv = 1.340781e+154                                                 fv = 11.39432                                                 fv = 6.198328                                                 fv = 10.31764                                                 fv = 4.597743                                                 fv = 5.873984                                                 fv = 432.4203                                                 fv = 1323.426                                                 fv = 6.082317                                                 fv = 6.044706                                                 fv = 4.597743                                                 fv = 6.408829                                                 fv = 1.340781e+154                                                 fv = 77.54922                                                 fv = 6.294637                                                 fv = 4.597743                                                 fv = 20.3289                                                 fv = 5.552039                                                 fv = 4.597743                                                 fv = 4.597743                                                 fv = 43.58494                                                 fv = 36.75275                                                 fv = 1.340781e+154                                                 fv = 13.54879                                                 fv = 6.703264                                                 fv = 5.081891                                                 fv = 4.616308                                                 fv = 11.43817                                                 fv = 8.758991                                                 fv = 5.420157                                                 fv = 4.986182                                                 fv = 5.095108                                                 fv = 5.000271                                                 fv = 5.13111                                                 fv = 4.597743                                                 fv = 4.597743                                                 fv = 1.340781e+154                                                 fv = 5.917191                                                 fv = 4.597743                                                 fv = 4.597743                                                 fv = 4.597743                                                 fv = 1.340781e+154                                                 fv = 13.70445                                                 fv = 6.445888                                                 fv = 13.70445 run # 1: f=4.597742547
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.08682                                                 fv = 11.13176                                                 fv = 11.13176                                                 fv = 11.26048                                                 fv = 360.8189                                                 fv = 11.13176                                                 fv = 1.340781e+154                                                 fv = 11.53322                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.384239                                                 fv = 101.6351                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14798969522                                                 fv = 8.384239                                                 fv = 15.95915                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.8857                                                 fv = 7644261110                                                 fv = 44229.98                                                 fv = 1.340781e+154                                                 fv = 6.328615                                                 fv = 6.328615                                                 fv = 1.340781e+154                                                 fv = 2187685577                                                 fv = 51.22715                                                 fv = 5.6376                                                 fv = 1.340781e+154                                                 fv = 8.846366                                                 fv = 1.340781e+154                                                 fv = 12378462                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.900332                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.02565                                                 fv = 10.79658                                                 fv = 12.02017                                                 fv = 12.02017                                                 fv = 5.6376                                                 fv = 10.79658                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.6376                                                 fv = 121.5964                                                 fv = 5.6376                                                 fv = 1.340781e+154                                                 fv = 11.01513                                                 fv = 11117.31                                                 fv = 87.06814                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.07384                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.29756                                                 fv = 5.6376                                                 fv = 1.340781e+154                                                 fv = 14.41831                                                 fv = 7002.249                                                 fv = 13.17276                                                 fv = 14.41831                                                 fv = 3.12698                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.45575                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.463915                                                 fv = 1.340781e+154                                                 fv = 7.14694                                                 fv = 1.340781e+154                                                 fv = 19.80608                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.12698                                                 fv = 23.70721                                                 fv = 1.340781e+154                                                 fv = 9.442581                                                 fv = 54.37261                                                 fv = 38.50831                                                 fv = 58.60728                                                 fv = 58.8536                                                 fv = 31.06025                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.47018                                                 fv = 3.12698                                                 fv = 780.9904                                                 fv = 11700733                                                 fv = 1001.972                                                 fv = 4.282658                                                 fv = 18.12805                                                 fv = 1.340781e+154                                                 fv = 6.331889                                                 fv = 1.340781e+154                                                 fv = 6.455789                                                 fv = 1.340781e+154                                                 fv = 3.662165                                                 fv = 1.340781e+154                                                 fv = 9.108037                                                 fv = 3.12698                                                 fv = 15.82211                                                 fv = 3.12698                                                 fv = 11.75802                                                 fv = 5.534396                                                 fv = 1.340781e+154                                                 fv = 7.795393                                                 fv = 22.4391                                                 fv = 31.1963                                                 fv = 12.93403                                                 fv = 3.12698                                                 fv = 13.98245                                                 fv = 291.0235                                                 fv = 15.55448                                                 fv = 1.340781e+154                                                 fv = 3.490732                                                 fv = 15.84844                                                 fv = 3.12698                                                 fv = 5.488223                                                 fv = 3.12698                                                 fv = 3.12698                                                 fv = 1.340781e+154                                                 fv = 17.47198                                                 fv = 3.600917                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=3.126979806
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53453.81                                                 fv = 36.75275                                                 fv = 36.75275                                                 fv = 36.75275                                                 fv = 36.75275                                                 fv = 36.75275                                                 fv = 12.12996                                                 fv = 7.183075                                                 fv = 1010.664                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.41058                                                 fv = 11.41058                                                 fv = 11.41058                                                 fv = 25694860371                                                 fv = 1.340781e+154                                                 fv = 17.95253                                                 fv = 21807337614                                                 fv = 8.306154                                                 fv = 92258.92                                                 fv = 9.157096                                                 fv = 7.272015                                                 fv = 14.44078                                                 fv = 14.44078                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.04538                                                 fv = 13.04538                                                 fv = 1.340781e+154                                                 fv = 5397.742                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2773.753                                                 fv = 2938.42                                                 fv = 144.8295                                                 fv = 8.020684                                                 fv = 9.157096                                                 fv = 1.340781e+154                                                 fv = 9.846748                                                 fv = 7.183075                                                 fv = 7.323715                                                 fv = 8.985384                                                 fv = 8.73601                                                 fv = 10.95074                                                 fv = 36929.46                                                 fv = 6.217139                                                 fv = 435.6885                                                 fv = 1.340781e+154                                                 fv = 9.688123                                                 fv = 34.2651                                                 fv = 1.340781e+154                                                 fv = 10.95074                                                 fv = 34.2651                                                 fv = 10.95074                                                 fv = 12.2206                                                 fv = 34.2651                                                 fv = 1.340781e+154                                                 fv = 11.54433                                                 fv = 59854400693                                                 fv = 11.85863                                                 fv = 1.340781e+154                                                 fv = 6.217139                                                 fv = 7.11587                                                 fv = 6.217139                                                 fv = 1.340781e+154                                                 fv = 1317.767                                                 fv = 6.444122                                                 fv = 6.592133                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.145991                                                 fv = 12.09328                                                 fv = 1.340781e+154                                                 fv = 346.562                                                 fv = 1.340781e+154                                                 fv = 12.09328                                                 fv = 10.10928                                                 fv = 7.223295                                                 fv = 7.01629                                                 fv = 6.998073                                                 fv = 2142.231                                                 fv = 7.223295                                                 fv = 6.892567                                                 fv = 8.792454                                                 fv = 7.945196                                                 fv = 7.833665                                                 fv = 6.190697                                                 fv = 1.340781e+154                                                 fv = 8.792454                                                 fv = 3326.716                                                 fv = 1.340781e+154                                                 fv = 6.190697                                                 fv = 12.49177                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 49.64332                                                 fv = 1.340781e+154                                                 fv = 14.1847                                                 fv = 7.323715                                                 fv = 6.01088                                                 fv = 397.3568                                                 fv = 7.183075                                                 fv = 1.340781e+154                                                 fv = 14.11432                                                 fv = 5.878845                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.878845                                                 fv = 6.145991                                                 fv = 6.145991                                                 fv = 12.09328                                                 fv = 19.5009                                                 fv = 5.878845                                                 fv = 12.9835                                                 fv = 6.536643                                                 fv = 6.629299                                                 fv = 6.629299                                                 fv = 143.3163                                                 fv = 8.368022                                                 fv = 8.620276                                                 fv = 9.92818                                                 fv = 1.340781e+154                                                 fv = 80.5338                                                 fv = 10.92801                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.878845                                                 fv = 12.36081                                                 fv = 3703.332                                                 fv = 6.629299                                                 fv = 6.960843                                                 fv = 5.878845                                                 fv = 5.878845                                                 fv = 9.807569                                                 fv = 5.878845                                                 fv = 5.878845                                                 fv = 2.05175                                                 fv = 4.359368                                                 fv = 1.340781e+154                                                 fv = 6.145991                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 2.101268                                                 fv = 3.996645                                                 fv = 1.711378                                                 fv = 1.340781e+154                                                 fv = 1.711378                                                 fv = 5.909936                                                 fv = 3.954893                                                 fv = 3.436767                                                 fv = 1.340781e+154                                                 fv = 12.16642                                                 fv = 1.83953                                                 fv = 9.846748                                                 fv = 12.17217                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.954893                                                 fv = 1.711378                                                 fv = 6611.07                                                 fv = 1.340781e+154                                                 fv = 12.72153                                                 fv = 7.289271                                                 fv = 2.529832                                                 fv = 1.340781e+154                                                 fv = 2.507194                                                 fv = 2.098547                                                 fv = 1.340781e+154                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 4.297242                                                 fv = 1.340781e+154                                                 fv = 1.711378                                                 fv = 1.987848                                                 fv = 1.340781e+154                                                 fv = 2.494696                                                 fv = 3.954893                                                 fv = 3.954893                                                 fv = 4.32504                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 291.0334                                                 fv = 14.37622                                                 fv = 10.88881                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 2.500204                                                 fv = 11.99828                                                 fv = 1.340781e+154                                                 fv = 2.330988                                                 fv = 1.926842                                                 fv = 2.05175                                                 fv = 1.83953                                                 fv = 2.101268                                                 fv = 2.101268                                                 fv = 6.130523                                                 fv = 9.966821                                                 fv = 14.37622                                                 fv = 1.945298                                                 fv = 5.272773                                                 fv = 1.711378                                                 fv = 3.996645                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 3.561255                                                 fv = 1.711378                                                 fv = 3.80115                                                 fv = 1.340781e+154                                                 fv = 1.925923                                                 fv = 20.5059                                                 fv = 1.340781e+154                                                 fv = 1.711378                                                 fv = 1.711378                                                 fv = 2.098547                                                 fv = 12.72153 run # 3: f=1.711377513
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 814
#> best    : 1.711377513
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 0 3 0 2 1 0 0 0 3 2 1 1 4 1 1 1 10 1 ) f(x) = 1.711377513
#> 
#>                                                 fv = 1.711378                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (5): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...          

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.95182