Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 8.873943                                                 fv = 15.87618                                                 fv = 11.59679                                                 fv = 181.2611                                                 fv = 44356.51                                                 fv = 16.02723                                                 fv = 826177.4                                                 fv = 2142.27                                                 fv = 16.06379                                                 fv = 73708.53                                                 fv = 1.340781e+154                                                 fv = 10.47529                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 64.69519                                                 fv = 1.340781e+154                                                 fv = 301.3989                                                 fv = 48398.68                                                 fv = 1.340781e+154                                                 fv = 1673.245                                                 fv = 28.02449                                                 fv = 1.340781e+154                                                 fv = 61.53                                                 fv = 1.340781e+154                                                 fv = 59.53363                                                 fv = 5737.739 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 51.2195                                                 fv = 8.182952                                                 fv = 26531.34                                                 fv = 398193.4                                                 fv = 1.340781e+154                                                 fv = 5.476405                                                 fv = 20.58638                                                 fv = 1.340781e+154                                                 fv = 24.23575                                                 fv = 16.96922                                                 fv = 6.871402                                                 fv = 14.98418                                                 fv = 57.28132                                                 fv = 8.337993                                                 fv = 6.544086                                                 fv = 13.56641                                                 fv = 8753.841                                                 fv = 21.7874                                                 fv = 1.340781e+154                                                 fv = 15.10181                                                 fv = 486.5652                                                 fv = 664.7727                                                 fv = 44.28294                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.38672                                                 fv = 1.340781e+154                                                 fv = 52.53493                                                 fv = 6.309331                                                 fv = 1863.287                                                 fv = 1.340781e+154                                                 fv = 258.2792                                                 fv = 11.06546                                                 fv = 8.299385                                                 fv = 8.381375                                                 fv = 31.21268                                                 fv = 9.292364                                                 fv = 8.674841                                                 fv = 5.476405                                                 fv = 9.170534                                                 fv = 6.544086                                                 fv = 17.05645                                                 fv = 7.57504                                                 fv = 8.337993                                                 fv = 6.250526                                                 fv = 1.340781e+154                                                 fv = 8.740839                                                 fv = 22.73346                                                 fv = 8.947753 run # 0: f=5.476404869
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.476496                                                 fv = 45.34965                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.688738                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.562277                                                 fv = 1.340781e+154                                                 fv = 20388733                                                 fv = 1.340781e+154                                                 fv = 19.16519                                                 fv = 291806070633                                                 fv = 105362.4                                                 fv = 1.340781e+154                                                 fv = 36636431508                                                 fv = 7.977433                                                 fv = 14.60419                                                 fv = 7.15113                                                 fv = 8.336523                                                 fv = 8.374845                                                 fv = 7.960025                                                 fv = 22.82037                                                 fv = 22.31781                                                 fv = 1.340781e+154                                                 fv = 8753889711                                                 fv = 143156.6                                                 fv = 24.25122                                                 fv = 1.721925e+13                                                 fv = 15.17411                                                 fv = 12.51585                                                 fv = 11.5768                                                 fv = 7.15113                                                 fv = 1.340781e+154                                                 fv = 15.17411                                                 fv = 16.38702                                                 fv = 651.934                                                 fv = 1.340781e+154                                                 fv = 1461.975                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.70969 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1200.599                                                 fv = 17.1116                                                 fv = 2853675                                                 fv = 11.39937                                                 fv = 8.866323                                                 fv = 7.15113                                                 fv = 8.249792                                                 fv = 14.77245                                                 fv = 7.15113                                                 fv = 11.95321                                                 fv = 7.15113                                                 fv = 8.866323                                                 fv = 1.340781e+154                                                 fv = 3084.267                                                 fv = 1.340781e+154                                                 fv = 7.907404                                                 fv = 9738.709                                                 fv = 12.79267                                                 fv = 8.292942                                                 fv = 1.340781e+154                                                 fv = 7.210865                                                 fv = 7.589776                                                 fv = 6.393721                                                 fv = 7416178                                                 fv = 1.340781e+154                                                 fv = 36.74597                                                 fv = 9.619709                                                 fv = 1.340781e+154                                                 fv = 9.454569                                                 fv = 6.393721                                                 fv = 6.393721                                                 fv = 12.04494                                                 fv = 12.67788                                                 fv = 6.393721                                                 fv = 1.340781e+154                                                 fv = 13.93852                                                 fv = 10.97558                                                 fv = 7.965774                                                 fv = 8.00399                                                 fv = 12.89071                                                 fv = 7.575118                                                 fv = 1.340781e+154                                                 fv = 22.42141                                                 fv = 6.775428                                                 fv = 7.353914                                                 fv = 6.393721                                                 fv = 6.393721                                                 fv = 6.67091                                                 fv = 12.67788                                                 fv = 9.454569                                                 fv = 7.832692                                                 fv = 1.340781e+154                                                 fv = 6.393721                                                 fv = 6.621115                                                 fv = 7.965774                                                 fv = 7.900618                                                 fv = 6.393721                                                 fv = 7.257883                                                 fv = 7.598621                                                 fv = 8.13866                                                 fv = 6.393721                                                 fv = 6.393721                                                 fv = 5.936789                                                 fv = 7.158519                                                 fv = 189.7052                                                 fv = 10.80822                                                 fv = 5.936789                                                 fv = 5.936789                                                 fv = 1.340781e+154                                                 fv = 6.903649                                                 fv = 25.54639                                                 fv = 5.936789                                                 fv = 7.08552                                                 fv = 6.704242                                                 fv = 5.936789                                                 fv = 5.936789                                                 fv = 7.364685                                                 fv = 6.822275                                                 fv = 6.220047                                                 fv = 7.939225                                                 fv = 7.920227                                                 fv = 1.340781e+154                                                 fv = 7.339777                                                 fv = 6.94257                                                 fv = 7.237299                                                 fv = 6.988157                                                 fv = 10.08446 run # 1: f=5.936788609
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.37494                                                 fv = 45.34965                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.16305                                                 fv = 15.38921                                                 fv = 23.61276                                                 fv = 12.77823                                                 fv = 22.27968                                                 fv = 48.52129                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.70508                                                 fv = 16.09885                                                 fv = 41.76387                                                 fv = 16.09885                                                 fv = 11.79898                                                 fv = 151.1339                                                 fv = 17.26379                                                 fv = 300.5236                                                 fv = 202.5893                                                 fv = 69.50559                                                 fv = 11.03047                                                 fv = 541967                                                 fv = 1.340781e+154                                                 fv = 15.315                                                 fv = 1.340781e+154                                                 fv = 2.286886e+12                                                 fv = 45.34965                                                 fv = 45.34965                                                 fv = 45.34965                                                 fv = 259.4562                                                 fv = 178.3605                                                 fv = 55.15236                                                 fv = 15.315                                                 fv = 381004.1                                                 fv = 1.340781e+154                                                 fv = 2.305092e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.03047                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.34965                                                 fv = 29.2177                                                 fv = 4551.929                                                 fv = 69.20657                                                 fv = 69.20657                                                 fv = 151.1339                                                 fv = 69.20657                                                 fv = 11.03047                                                 fv = 1.340781e+154                                                 fv = 15.315                                                 fv = 14.26612                                                 fv = 1.340781e+154                                                 fv = 11.03047                                                 fv = 151.1339                                                 fv = 40718.75                                                 fv = 12922.94                                                 fv = 17.78254                                                 fv = 51.76013                                                 fv = 36.74597                                                 fv = 1.340781e+154                                                 fv = 43.30171                                                 fv = 7.444208                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.813038                                                 fv = 28104.45                                                 fv = 13.50401                                                 fv = 15.58129                                                 fv = 9.51499                                                 fv = 1.340781e+154                                                 fv = 16.26173                                                 fv = 7.444208                                                 fv = 7.444208                                                 fv = 1.340781e+154                                                 fv = 74.86078                                                 fv = 7.444208                                                 fv = 7.444208                                                 fv = 8.630077                                                 fv = 8.630077                                                 fv = 7.444208                                                 fv = 7.444208                                                 fv = 10.11105                                                 fv = 22.72018                                                 fv = 1.340781e+154                                                 fv = 12.92737                                                 fv = 8.630077                                                 fv = 7.444208                                                 fv = 12.24032                                                 fv = 7.444208                                                 fv = 7.444208                                                 fv = 7.567596                                                 fv = 8.106759                                                 fv = 7.883653                                                 fv = 1.340781e+154                                                 fv = 7.444208                                                 fv = 8.21432                                                 fv = 7.444208                                                 fv = 17.03885                                                 fv = 7.444208                                                 fv = 1.340781e+154                                                 fv = 58.21872                                                 fv = 7.444208                                                 fv = 1.340781e+154 run # 2: f=7.444207957
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 279023.6                                                 fv = 45.34965                                                 fv = 29.94825                                                 fv = 9.801994                                                 fv = 80.00159                                                 fv = 28.64925                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5500.088                                                 fv = 22.2187                                                 fv = 22.2187                                                 fv = 36.74597                                                 fv = 1.93735e+12                                                 fv = 70580116156                                                 fv = 1.340781e+154                                                 fv = 118510.7                                                 fv = 8.368182                                                 fv = 24.70573                                                 fv = 8.507845                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.07041                                                 fv = 12218.67                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22192376                                                 fv = 331901873723                                                 fv = 8.368182                                                 fv = 2594.179                                                 fv = 73.5338                                                 fv = 34.45031                                                 fv = 1.340781e+154                                                 fv = 45.34965                                                 fv = 1.340781e+154                                                 fv = 17220202058                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.87456                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 35.81173                                                 fv = 121.1118                                                 fv = 36369.84                                                 fv = 1.340781e+154                                                 fv = 7.588145                                                 fv = 7.588145                                                 fv = 1.340781e+154                                                 fv = 40.37255                                                 fv = 1.340781e+154                                                 fv = 365324                                                 fv = 1.340781e+154                                                 fv = 38.74192                                                 fv = 8.790693                                                 fv = 7.588145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.85108                                                 fv = 1.340781e+154                                                 fv = 7.588145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.588145                                                 fv = 1.340781e+154                                                 fv = 3.473306e+12                                                 fv = 9480249181                                                 fv = 1.340781e+154                                                 fv = 22.45214                                                 fv = 1.340781e+154                                                 fv = 1280.238                                                 fv = 7.588145                                                 fv = 1.340781e+154                                                 fv = 1867.201                                                 fv = 20.47914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42.96635                                                 fv = 13.75424                                                 fv = 2969.882                                                 fv = 23.80941                                                 fv = 1.340781e+154                                                 fv = 7.588145                                                 fv = 41613.73                                                 fv = 6.878976                                                 fv = 7.303105                                                 fv = 1.340781e+154                                                 fv = 1474.971                                                 fv = 9.418731                                                 fv = 29.94825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.932219                                                 fv = 5.932219                                                 fv = 1.340781e+154                                                 fv = 32.73469                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21994.6                                                 fv = 6991692                                                 fv = 6.89153                                                 fv = 5.932219                                                 fv = 551.8165                                                 fv = 7.394097                                                 fv = 5.932219                                                 fv = 6.560215                                                 fv = 5088.15                                                 fv = 81.94889                                                 fv = 21921.16                                                 fv = 134.9588                                                 fv = 1244.691                                                 fv = 1.340781e+154                                                 fv = 8.052194                                                 fv = 6.480557                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.039725                                                 fv = 54.28768                                                 fv = 1.340781e+154                                                 fv = 303.6057                                                 fv = 9.998521                                                 fv = 14.61714                                                 fv = 11.23553                                                 fv = 1.340781e+154                                                 fv = 5.932219                                                 fv = 5.954507                                                 fv = 5.932219                                                 fv = 1.340781e+154                                                 fv = 5.932219                                                 fv = 5.932219                                                 fv = 10190.49                                                 fv = 81.94889                                                 fv = 132.4776                                                 fv = 50968190                                                 fv = 1.340781e+154                                                 fv = 29.94825                                                 fv = 1.340781e+154                                                 fv = 10.561                                                 fv = 1.340781e+154                                                 fv = 805.3929                                                 fv = 5.726885                                                 fv = 10.03764                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.726885                                                 fv = 6.870977                                                 fv = 6.710182                                                 fv = 5.726885                                                 fv = 5.726885                                                 fv = 5.726885                                                 fv = 1.340781e+154                                                 fv = 7.28021                                                 fv = 5.645127                                                 fv = 1.340781e+154                                                 fv = 10.561                                                 fv = 6.190804                                                 fv = 1833.467                                                 fv = 5.645127                                                 fv = 1.340781e+154                                                 fv = 5.906158                                                 fv = 5.645127                                                 fv = 6.568312                                                 fv = 5.645127                                                 fv = 5.645127                                                 fv = 5.976971                                                 fv = 1.340781e+154                                                 fv = 12.67988                                                 fv = 5.726885                                                 fv = 7.031511                                                 fv = 5.645127                                                 fv = 6.498943                                                 fv = 6.760338                                                 fv = 7.715928                                                 fv = 11.51322                                                 fv = 6.870977                                                 fv = 7.445333 run # 3: f=5.645127298
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 627
#> best    : 5.476404869
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 1 3 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 ) f(x) = 5.476404869
#> 
#>                                                 fv = 5.476405                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.774871