Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 23.12595                                                 fv = 420.706                                                 fv = 10.66304                                                 fv = 130522.1                                                 fv = 1.340781e+154                                                 fv = 14.30477                                                 fv = 17.9499                                                 fv = 39646268                                                 fv = 1.340781e+154                                                 fv = 26.67063                                                 fv = 41.84487                                                 fv = 1.340781e+154                                                 fv = 124.292                                                 fv = 1.340781e+154                                                 fv = 101.7735                                                 fv = 1.340781e+154                                                 fv = 17.9499                                                 fv = 17.9499                                                 fv = 1.340781e+154                                                 fv = 6466.689                                                 fv = 1.340781e+154                                                 fv = 48.82794                                                 fv = 320377136                                                 fv = 10651.32                                                 fv = 10.66304                                                 fv = 545986                                                 fv = 13.18705                                                 fv = 13.18705                                                 fv = 28.15829                                                 fv = 12.56679                                                 fv = 17.23295                                                 fv = 17.23295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.66915                                                 fv = 14.04106                                                 fv = 27.706                                                 fv = 20.79165                                                 fv = 7.248594                                                 fv = 1767029                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 854.8388                                                 fv = 4931.154                                                 fv = 7.248594                                                 fv = 43.68137                                                 fv = 7.248594                                                 fv = 7.498801                                                 fv = 6.106657                                                 fv = 1.340781e+154                                                 fv = 30.48254                                                 fv = 27.71409                                                 fv = 18.37524                                                 fv = 1.340781e+154                                                 fv = 16.78125                                                 fv = 13.68397                                                 fv = 20.51723                                                 fv = 13.68397                                                 fv = 34.03975                                                 fv = 45.54056                                                 fv = 27.58718                                                 fv = 6.106657                                                 fv = 12.16531                                                 fv = 1357.185                                                 fv = 57.56099                                                 fv = 5142.599                                                 fv = 7237.686                                                 fv = 101215.1                                                 fv = 260.2782                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 84.23364                                                 fv = 23.85611                                                 fv = 52.00715                                                 fv = 53.27847                                                 fv = 52.00715                                                 fv = 31.17456                                                 fv = 69.86056                                                 fv = 6.106657                                                 fv = 6.106657                                                 fv = 1.340781e+154                                                 fv = 96.18016                                                 fv = 6.106657                                                 fv = 1.340781e+154                                                 fv = 238.2121                                                 fv = 83.39442                                                 fv = 64.2653                                                 fv = 38.09973                                                 fv = 45.50099                                                 fv = 352.8881                                                 fv = 8381062                                                 fv = 11.0753                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.32482                                                 fv = 7.470561                                                 fv = 18.94929                                                 fv = 9.738962                                                 fv = 32.92453                                                 fv = 6.106657                                                 fv = 18.21003                                                 fv = 40.0911                                                 fv = 6.106657                                                 fv = 1843.81                                                 fv = 13.5694 run # 0: f=6.106657265
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.51666                                                 fv = 38.6039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.10108                                                 fv = 1.340781e+154                                                 fv = 9.354697                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.64318                                                 fv = 18.04636                                                 fv = 1.340781e+154                                                 fv = 47559                                                 fv = 4906608                                                 fv = 8.330493                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.681                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11757157049                                                 fv = 1.340781e+154                                                 fv = 297582133208                                                 fv = 4738.776                                                 fv = 8.6465                                                 fv = 1021339401                                                 fv = 10.1039                                                 fv = 8.330493                                                 fv = 1.340781e+154                                                 fv = 585695.8                                                 fv = 8.330493                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9458388538                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1141.056                                                 fv = 1.340781e+154                                                 fv = 108.5575                                                 fv = 1.340781e+154                                                 fv = 204.4185                                                 fv = 26.17558                                                 fv = 158687.4                                                 fv = 264.6918                                                 fv = 6.849388                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.37648                                                 fv = 6.849388                                                 fv = 19.11878                                                 fv = 11.90964                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.446365                                                 fv = 12.96932                                                 fv = 1.340781e+154                                                 fv = 6.849388                                                 fv = 26.82811                                                 fv = 197.5776                                                 fv = 23.58272                                                 fv = 1.340781e+154                                                 fv = 6.849388                                                 fv = 12.99806                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.87171                                                 fv = 14.33367                                                 fv = 1.340781e+154                                                 fv = 7.841892                                                 fv = 9.477042                                                 fv = 16.10021                                                 fv = 10.89703                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 17.15128                                                 fv = 6.849388 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.606353                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 278.3327                                                 fv = 21.03958                                                 fv = 9.585145                                                 fv = 1.340781e+154                                                 fv = 7.473446                                                 fv = 7.849379                                                 fv = 8.018211                                                 fv = 6.887632                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 6.849388                                                 fv = 9.938478                                                 fv = 6.921787                                                 fv = 6.849388                                                 fv = 1.340781e+154                                                 fv = 8.322219                                                 fv = 7.938858                                                 fv = 9.368376 run # 1: f=6.849388044
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.15439                                                 fv = 38.6039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.963421                                                 fv = 14.02272                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.14588                                                 fv = 9.142665                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.142665                                                 fv = 117.8937                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.66248                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 67121325                                                 fv = 9.142665                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 585918.1                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.42836                                                 fv = 9.927211                                                 fv = 28.00312                                                 fv = 9.295759                                                 fv = 9.142665                                                 fv = 1.340781e+154                                                 fv = 10.74244                                                 fv = 16861.68                                                 fv = 10.0316                                                 fv = 9.142665                                                 fv = 13.34184                                                 fv = 1585.347                                                 fv = 1.340781e+154                                                 fv = 14.15929                                                 fv = 1.340781e+154                                                 fv = 11.28297                                                 fv = 30.11183                                                 fv = 29.87652                                                 fv = 11.28297                                                 fv = 58188.5                                                 fv = 1.340781e+154                                                 fv = 73.2531                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.33281                                                 fv = 9.120384                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.33281                                                 fv = 9.120384                                                 fv = 1.340781e+154                                                 fv = 9.120384                                                 fv = 9.120384                                                 fv = 9.120384                                                 fv = 1.340781e+154                                                 fv = 9.120384                                                 fv = 8.786587                                                 fv = 1.340781e+154                                                 fv = 28.00312                                                 fv = 206.6259                                                 fv = 1.340781e+154                                                 fv = 3.071618                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.00312                                                 fv = 3.071618                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5373.952                                                 fv = 88913.45                                                 fv = 3.071618                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.20429                                                 fv = 23.43517                                                 fv = 253.9933                                                 fv = 3.071618                                                 fv = 37.56171                                                 fv = 33.91937                                                 fv = 19.82443                                                 fv = 23.43517                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.96323                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.07641                                                 fv = 3.071618                                                 fv = 1.340781e+154                                                 fv = 6.89257                                                 fv = 3.071618                                                 fv = 11.06859                                                 fv = 3.071618                                                 fv = 17.57171                                                 fv = 1.340781e+154                                                 fv = 11.77016                                                 fv = 3.071618                                                 fv = 145577.7                                                 fv = 894.8084                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.88374                                                 fv = 23.44084                                                 fv = 1.340781e+154                                                 fv = 50.23737                                                 fv = 1.340781e+154                                                 fv = 19.78857                                                 fv = 11.71659                                                 fv = 3.071618                                                 fv = 2.051274                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 54.32978                                                 fv = 8.634307                                                 fv = 4606012                                                 fv = 11.64302                                                 fv = 2.051274                                                 fv = 2.051274                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.051274                                                 fv = 65.95347                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.051274                                                 fv = 1.340781e+154                                                 fv = 2.401444                                                 fv = 2.051274                                                 fv = 0.766162                                                 fv = 305.538                                                 fv = 40.62058                                                 fv = 21.90732                                                 fv = 29.39551                                                 fv = 12.95261                                                 fv = 27.09684                                                 fv = 20.62395                                                 fv = 11.77016                                                 fv = 7469.035                                                 fv = 15.86091                                                 fv = 1.340781e+154                                                 fv = 0.766162                                                 fv = 0.766162                                                 fv = 0.766162                                                 fv = 1036.011                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.70186                                                 fv = 26.4785                                                 fv = 0.766162                                                 fv = 0.766162                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 130.305                                                 fv = 4496.041                                                 fv = 14.6121                                                 fv = 366.4987                                                 fv = 217.2284                                                 fv = 101.545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3600.971 run # 2: f=0.7661620442
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.89506                                                 fv = 38.6039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.89506                                                 fv = 14.19028                                                 fv = 1.340781e+154                                                 fv = 8.112138                                                 fv = 16899855                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3746.285                                                 fv = 937647.1                                                 fv = 1.340781e+154                                                 fv = 43731.54                                                 fv = 1.340781e+154                                                 fv = 9.190978                                                 fv = 1.340781e+154                                                 fv = 5.317917                                                 fv = 23.30833                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23852.97                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8197.988                                                 fv = 1375.267                                                 fv = 5.317917                                                 fv = 5.317917                                                 fv = 5738.563                                                 fv = 271.855                                                 fv = 1.340781e+154                                                 fv = 905.1966                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.31011                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.434725                                                 fv = 2.834433                                                 fv = 37751.97                                                 fv = 1.340781e+154                                                 fv = 10.83532                                                 fv = 3049.72                                                 fv = 5.85077                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.57647                                                 fv = 5.85077                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.834433                                                 fv = 482066.7                                                 fv = 1127651855                                                 fv = 485.6657                                                 fv = 1.340781e+154                                                 fv = 86.78596                                                 fv = 14397.26                                                 fv = 1.340781e+154                                                 fv = 92018.1                                                 fv = 14.61199                                                 fv = 5.85077                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.52556                                                 fv = 1.340781e+154                                                 fv = 2.834433                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.147385                                                 fv = 7129.724                                                 fv = 1.340781e+154                                                 fv = 2.834433                                                 fv = 14.06521                                                 fv = 2.834433                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.97633                                                 fv = 29.04963                                                 fv = 1.340781e+154                                                 fv = 59.41874                                                 fv = 2.834433                                                 fv = 9.550221                                                 fv = 12.09435                                                 fv = 1.340781e+154                                                 fv = 9.436214                                                 fv = 4.293125                                                 fv = 5.85077                                                 fv = 8.860105                                                 fv = 5.375509                                                 fv = 5.782018                                                 fv = 2.560961                                                 fv = 4966.311                                                 fv = 1.340781e+154                                                 fv = 6723.475                                                 fv = 9.280152                                                 fv = 1.340781e+154                                                 fv = 5.770073                                                 fv = 1.340781e+154                                                 fv = 11.8175                                                 fv = 1.340781e+154                                                 fv = 17.39984                                                 fv = 2.637052                                                 fv = 12.94677                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.560961                                                 fv = 2.560961                                                 fv = 4.658102                                                 fv = 4.22144                                                 fv = 5.510572                                                 fv = 5.948036                                                 fv = 5.347211                                                 fv = 6.406294                                                 fv = 2.560961                                                 fv = 2.560961                                                 fv = 34.86037                                                 fv = 11.59949                                                 fv = 325424.1 run # 4: f=2.560960796
#> 
#> bb eval : 620
#> best    : 0.7661620442
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 1 2 2 2 0 0 0 3 7 3 1 4 2 1 9 1 3 2 ) f(x) = 0.7661620442
#> 
#>                                                 fv = 0.766162                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (5): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 331.5351