Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 4.823285                                                 fv = 47.86701                                                 fv = 15.45701                                                 fv = 64504.79                                                 fv = 470.6045                                                 fv = 11.79432                                                 fv = 1884817                                                 fv = 9876.852                                                 fv = 6.569743                                                 fv = 60943.65                                                 fv = 1.340781e+154                                                 fv = 8.757275                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 145.3729                                                 fv = 1.340781e+154                                                 fv = 266.0047                                                 fv = 12263.3                                                 fv = 1.340781e+154                                                 fv = 1167.891                                                 fv = 15.82683                                                 fv = 1.340781e+154                                                 fv = 340.1766                                                 fv = 1.340781e+154                                                 fv = 135.6553                                                 fv = 826.131 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 43.49288                                                 fv = 7.583633                                                 fv = 17.07379                                                 fv = 21.01632                                                 fv = 31.967                                                 fv = 17.96153                                                 fv = 21.99686                                                 fv = 1.340781e+154                                                 fv = 29.19333                                                 fv = 1.340781e+154                                                 fv = 25.15828                                                 fv = 10.09268                                                 fv = 1.340781e+154                                                 fv = 15.15942                                                 fv = 8.887296                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.08869 run # 0: f=4.823284735
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.09466                                                 fv = 19.96215                                                 fv = 11.13443                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.131667                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.36121                                                 fv = 1.340781e+154                                                 fv = 8672855                                                 fv = 2101450939                                                 fv = 7.489271                                                 fv = 1.340781e+154                                                 fv = 125626.3                                                 fv = 1.340781e+154                                                 fv = 16.37655                                                 fv = 7.489271                                                 fv = 1.340781e+154                                                 fv = 2.416341e+14                                                 fv = 1.340781e+154                                                 fv = 456.8105                                                 fv = 1.401248e+12                                                 fv = 1.340781e+154                                                 fv = 16.37655                                                 fv = 1.340781e+154                                                 fv = 516070853709                                                 fv = 17.47976                                                 fv = 456.8105                                                 fv = 5678.545                                                 fv = 7.074898                                                 fv = 5.954918                                                 fv = 5.954918                                                 fv = 1.340781e+154                                                 fv = 5.555029                                                 fv = 5.555029                                                 fv = 1411.125                                                 fv = 5.881425                                                 fv = 5.881425                                                 fv = 5.881425                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.31774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.95001                                                 fv = 1.340781e+154                                                 fv = 888.1462                                                 fv = 11.84319                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 1.340781e+154                                                 fv = 1139.594                                                 fv = 6.081874                                                 fv = 6.081874                                                 fv = 6.081874                                                 fv = 1.340781e+154                                                 fv = 5.778418                                                 fv = 1.340781e+154                                                 fv = 7.107486                                                 fv = 5.011918                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.635862                                                 fv = 5.305477                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1282.272                                                 fv = 1.340781e+154                                                 fv = 5.305477                                                 fv = 4042532                                                 fv = 1.340781e+154                                                 fv = 41712270                                                 fv = 1.105209e+13                                                 fv = 1.340781e+154                                                 fv = 10.37939                                                 fv = 1.340781e+154                                                 fv = 6.028283                                                 fv = 7.521103                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 5.011918                                                 fv = 5.153592                                                 fv = 5.701132                                                 fv = 5.011918                                                 fv = 70.58156                                                 fv = 16.36331                                                 fv = 19.37614                                                 fv = 78466.94                                                 fv = 14.50804                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 5.011918                                                 fv = 80.04555                                                 fv = 1.340781e+154                                                 fv = 7.417211                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8562.859                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.189232                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 5.011918                                                 fv = 5.011918                                                 fv = 4.992992                                                 fv = 9.416955                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 4.992992                                                 fv = 1.340781e+154                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 5.692067                                                 fv = 4.992992                                                 fv = 1.340781e+154                                                 fv = 288.7599                                                 fv = 5.894472                                                 fv = 5.05771                                                 fv = 13.05688                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 15.68141                                                 fv = 9.299353                                                 fv = 5.298084                                                 fv = 4.992992                                                 fv = 5.905491                                                 fv = 14.46625                                                 fv = 1.340781e+154                                                 fv = 5.011918                                                 fv = 6.181323                                                 fv = 6.770038                                                 fv = 1.340781e+154                                                 fv = 8.247015                                                 fv = 5.932239                                                 fv = 1.340781e+154                                                 fv = 8.358748                                                 fv = 6.463588                                                 fv = 14.4016                                                 fv = 4.992992                                                 fv = 5.316247                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 4.992992                                                 fv = 19.96215 run # 1: f=4.992991569
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 265.3666                                                 fv = 19.96215                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 978721.4                                                 fv = 19.96215                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.96215                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.239747                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.012571e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.73585e+13                                                 fv = 156057241385                                                 fv = 2.73585e+13                                                 fv = 2.73585e+13                                                 fv = 836930512                                                 fv = 1.340781e+154                                                 fv = 26.6474                                                 fv = 1.340781e+154                                                 fv = 24.56696                                                 fv = 16.96769                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.691459e+14                                                 fv = 7.396604e+13                                                 fv = 9402373510                                                 fv = 9.093146e+12                                                 fv = 9.093146e+12                                                 fv = 19.06501                                                 fv = 8.239747                                                 fv = 8.239747                                                 fv = 8.239747                                                 fv = 1.340781e+154                                                 fv = 8.239747                                                 fv = 1.340781e+154                                                 fv = 7.749256e+13                                                 fv = 7.75896e+13                                                 fv = 43.03367                                                 fv = 19.37614                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.401326                                                 fv = 7.834226                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.819999                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.032085                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.84628                                                 fv = 10.58969                                                 fv = 118.3018                                                 fv = 8.28528                                                 fv = 548.2529                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.401326                                                 fv = 7.401326                                                 fv = 7.401326                                                 fv = 7.401326                                                 fv = 16.62572                                                 fv = 18.52811                                                 fv = 520355364448                                                 fv = 11.29424                                                 fv = 8.423692                                                 fv = 9.303285                                                 fv = 5.638801                                                 fv = 536.3933                                                 fv = 1.340781e+154                                                 fv = 6.287943                                                 fv = 173.412                                                 fv = 1.340781e+154                                                 fv = 7.215271                                                 fv = 7.692816                                                 fv = 5.638801                                                 fv = 1.340781e+154                                                 fv = 6.316472                                                 fv = 5.961779                                                 fv = 5.638801                                                 fv = 5.638801                                                 fv = 5.785117                                                 fv = 20.25904                                                 fv = 5.638801                                                 fv = 3402.459                                                 fv = 1.340781e+154                                                 fv = 17.00618                                                 fv = 7.378275                                                 fv = 5.090734                                                 fv = 13.21571                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 73.54691                                                 fv = 1.340781e+154                                                 fv = 11.76444                                                 fv = 5.638801                                                 fv = 5.090734                                                 fv = 28.21938                                                 fv = 4.765052                                                 fv = 49312.58                                                 fv = 1.340781e+154                                                 fv = 5.56323                                                 fv = 25.00173                                                 fv = 11.33224                                                 fv = 1.340781e+154                                                 fv = 11.16283                                                 fv = 1.340781e+154                                                 fv = 53.91589                                                 fv = 4.765052                                                 fv = 4.765052                                                 fv = 9.036312                                                 fv = 4.765052                                                 fv = 4.765052                                                 fv = 7.245192                                                 fv = 4.826114                                                 fv = 15.50334                                                 fv = 12.81975                                                 fv = 5.816212                                                 fv = 12.81975                                                 fv = 12.81975                                                 fv = 5.660719                                                 fv = 1.340781e+154                                                 fv = 10.80822                                                 fv = 10.74597                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.703633                                                 fv = 69.82005                                                 fv = 9.271412                                                 fv = 8.494537                                                 fv = 11.37564                                                 fv = 1.340781e+154                                                 fv = 8.494537 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 20.98642                                                 fv = 1.340781e+154                                                 fv = 8.494537                                                 fv = 8.494537                                                 fv = 18.85454                                                 fv = 8.546509                                                 fv = 4.765052                                                 fv = 4.765052                                                 fv = 4.765052                                                 fv = 5.350799 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 12.9772                                                 fv = 11.29424                                                 fv = 171.1753                                                 fv = 1.340781e+154                                                 fv = 8.153                                                 fv = 1.340781e+154                                                 fv = 4.497783                                                 fv = 4856.99                                                 fv = 72.11127                                                 fv = 4.308956                                                 fv = 1.340781e+154                                                 fv = 5.785117                                                 fv = 4.497783                                                 fv = 4.54371                                                 fv = 7.167178                                                 fv = 4.308956                                                 fv = 4.242767                                                 fv = 5.676983                                                 fv = 1.340781e+154                                                 fv = 5.449679                                                 fv = 6.431881                                                 fv = 6.424715                                                 fv = 1.340781e+154                                                 fv = 2.888192                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.54219                                                 fv = 6.203013                                                 fv = 3.765528                                                 fv = 4.723384                                                 fv = 2.888192                                                 fv = 1.340781e+154                                                 fv = 2.564699                                                 fv = 1.340781e+154                                                 fv = 9.933299                                                 fv = 3.729925                                                 fv = 2.829574                                                 fv = 2.564699                                                 fv = 4.070463                                                 fv = 3.870716                                                 fv = 2.564699                                                 fv = 4.371176                                                 fv = 2.940917                                                 fv = 1.340781e+154                                                 fv = 3.046452                                                 fv = 2.564699                                                 fv = 2.564699                                                 fv = 15.87742                                                 fv = 4.051011                                                 fv = 3.749644                                                 fv = 5.471414                                                 fv = 2.328489                                                 fv = 1.340781e+154                                                 fv = 16.07091                                                 fv = 1.340781e+154                                                 fv = 3.857376                                                 fv = 2.551456                                                 fv = 3.521121                                                 fv = 11.0805                                                 fv = 8.613828                                                 fv = 3.194076                                                 fv = 3.801357                                                 fv = 12.0844                                                 fv = 3.478486                                                 fv = 3.708588                                                 fv = 2.328489                                                 fv = 2.328489                                                 fv = 3.000679                                                 fv = 2.944512                                                 fv = 4.205512                                                 fv = 2.358723                                                 fv = 2.564699                                                 fv = 4.64874                                                 fv = 3.305518                                                 fv = 6.732137                                                 fv = 1.340781e+154                                                 fv = 2.328489                                                 fv = 2.328489                                                 fv = 5.143487                                                 fv = 14.54318                                                 fv = 8.366212                                                 fv = 7.196296                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.11739                                                 fv = 9.877203                                                 fv = 2.328489                                                 fv = 2.551456                                                 fv = 3.708588                                                 fv = 3.521121                                                 fv = 7.367291                                                 fv = 19.29036                                                 fv = 1.340781e+154                                                 fv = 4.806299                                                 fv = 2.822757                                                 fv = 48.5044                                                 fv = 1.340781e+154                                                 fv = 2.70337                                                 fv = 13.82045                                                 fv = 8.927249 run # 2: f=2.328489196
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.68626                                                 fv = 19.96215                                                 fv = 8.239747                                                 fv = 8.197086                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.685416                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.471                                                 fv = 8.786035                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 10.3381                                                 fv = 1.340781e+154                                                 fv = 524.0918                                                 fv = 9.474433                                                 fv = 296.5888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 356.8866                                                 fv = 1.340781e+154                                                 fv = 10.3381                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 6.696172                                                 fv = 14.95802                                                 fv = 13.7665                                                 fv = 9.710872                                                 fv = 78.16258                                                 fv = 4.791401                                                 fv = 799265.1                                                 fv = 1.340781e+154                                                 fv = 20.64572                                                 fv = 2.097957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.64572                                                 fv = 2.097957                                                 fv = 2.097957                                                 fv = 2.097957                                                 fv = 1.340781e+154                                                 fv = 48.27719                                                 fv = 19.74551                                                 fv = 1.340781e+154                                                 fv = 2.097957                                                 fv = 22.11393                                                 fv = 1.453388                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 1.453388                                                 fv = 1.340781e+154                                                 fv = 84.37355                                                 fv = 1.453388                                                 fv = 5.718843                                                 fv = 11.79122                                                 fv = 25.76524                                                 fv = 11.24796                                                 fv = 1.32323                                                 fv = 17.63024                                                 fv = 1.340781e+154                                                 fv = 6.696172                                                 fv = 1.32323                                                 fv = 8.118769                                                 fv = 1.340781e+154                                                 fv = 41.15356                                                 fv = 2.141044                                                 fv = 17.63024                                                 fv = 11.97471                                                 fv = 14.02419                                                 fv = 19.39522                                                 fv = 21.68781                                                 fv = 15.77661                                                 fv = 1.340781e+154                                                 fv = 60.90263                                                 fv = 18.61679                                                 fv = 21.28828                                                 fv = 430.0166                                                 fv = 1.32323                                                 fv = 2216.564                                                 fv = 17.63024                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.32323                                                 fv = 1.32323                                                 fv = 3.25713                                                 fv = 7.924336                                                 fv = 1.32323                                                 fv = 3.376275                                                 fv = 1.340647                                                 fv = 1.340781e+154                                                 fv = 1.32323                                                 fv = 17.63024                                                 fv = 1.453388                                                 fv = 13.3559                                                 fv = 1.717447                                                 fv = 1.340781e+154                                                 fv = 20.67568                                                 fv = 4.109727                                                 fv = 2.113099                                                 fv = 6.974788                                                 fv = 2.649362                                                 fv = 8.364607                                                 fv = 62.58991                                                 fv = 1.32323                                                 fv = 400.1009                                                 fv = 92.29527                                                 fv = 1.340781e+154                                                 fv = 7.685416                                                 fv = 25.91286                                                 fv = 2.010721                                                 fv = 1.340781e+154                                                 fv = 1.32323                                                 fv = 1.32323                                                 fv = 13.08743                                                 fv = 1.340781e+154                                                 fv = 1.32323                                                 fv = 6.974788                                                 fv = 78.53602                                                 fv = 2.211487                                                 fv = 38.16988 run # 4: f=1.32323031
#> 
#> bb eval : 767
#> best    : 1.32323031
#> worst   : 1.340780793e+154
#> solution: x = ( 0 4 2 0 0 0 2 1 0 0 1 1 1 2 1 1 3 1 1 1 ) f(x) = 1.32323031
#> 
#>                                                 fv = 1.32323                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 62.83352