Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 19.94283                                                 fv = 27.74314                                                 fv = 12.64702                                                 fv = 2939183                                                 fv = 1.340781e+154                                                 fv = 16.32291                                                 fv = 24.52456                                                 fv = 58341742                                                 fv = 1.340781e+154                                                 fv = 43.17243                                                 fv = 78.97714                                                 fv = 1.340781e+154                                                 fv = 78.87894                                                 fv = 1.340781e+154                                                 fv = 4618.465                                                 fv = 1.340781e+154                                                 fv = 24.52456                                                 fv = 24.52456                                                 fv = 1.340781e+154                                                 fv = 310.5235                                                 fv = 1.340781e+154                                                 fv = 223.3136                                                 fv = 81067752                                                 fv = 955.8195                                                 fv = 28342.4                                                 fv = 9.971548                                                 fv = 15.99276                                                 fv = 42.32209                                                 fv = 13.82897                                                 fv = 1199.854                                                 fv = 83.48224                                                 fv = 15.73441                                                 fv = 83.48224                                                 fv = 1.340781e+154                                                 fv = 10.2238                                                 fv = 1.340781e+154                                                 fv = 9.971548                                                 fv = 616.3005                                                 fv = 21.17363                                                 fv = 14.35046                                                 fv = 11.44605                                                 fv = 26.10361                                                 fv = 14.74316 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 15.73436                                                 fv = 10.53254                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.2478                                                 fv = 39.5597                                                 fv = 1.340781e+154                                                 fv = 32.70104                                                 fv = 34.44042                                                 fv = 16.8745                                                 fv = 1.340781e+154                                                 fv = 25.76763                                                 fv = 12.64702                                                 fv = 1.340781e+154                                                 fv = 9.971548                                                 fv = 9.971548                                                 fv = 1.340781e+154                                                 fv = 13.82897                                                 fv = 61.63981                                                 fv = 15.98645                                                 fv = 9.971548                                                 fv = 40.50696                                                 fv = 14.19254                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.09172                                                 fv = 17.48331                                                 fv = 25.18193                                                 fv = 18.4311                                                 fv = 13.33467                                                 fv = 1.340781e+154                                                 fv = 9.971548                                                 fv = 1.340781e+154                                                 fv = 9.971548                                                 fv = 9.971548                                                 fv = 52.1383                                                 fv = 10.81209                                                 fv = 10.99574                                                 fv = 27.31081                                                 fv = 1.340781e+154 run # 0: f=9.971547804
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.05908                                                 fv = 43.34765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.05908                                                 fv = 26.66489                                                 fv = 2834.44                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 617077370                                                 fv = 1.340781e+154                                                 fv = 40.09084                                                 fv = 19.81373                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.0095                                                 fv = 1.340781e+154                                                 fv = 75601.85                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 183199.4                                                 fv = 1.340781e+154                                                 fv = 19.81373                                                 fv = 7319.993                                                 fv = 24.67767                                                 fv = 11.51585                                                 fv = 30.99274                                                 fv = 7.396523                                                 fv = 1964.218                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3424.529                                                 fv = 35.78448                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4188.032                                                 fv = 1.340781e+154                                                 fv = 82.82011                                                 fv = 16654.86                                                 fv = 5852433                                                 fv = 8.126563e+12                                                 fv = 1.340781e+154                                                 fv = 5592.013                                                 fv = 4188.032                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2813162693                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.46365                                                 fv = 13.28071                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.58262                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.518343                                                 fv = 6.375017                                                 fv = 6.375017                                                 fv = 1.340781e+154                                                 fv = 30.97669                                                 fv = 5040717                                                 fv = 1.340781e+154                                                 fv = 7.809699                                                 fv = 1.340781e+154                                                 fv = 405303469                                                 fv = 13.27073                                                 fv = 1.340781e+154                                                 fv = 1307.5                                                 fv = 1.340781e+154                                                 fv = 10.42153 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.375017                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3241745503                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 787.4622                                                 fv = 6.718339                                                 fv = 6.103554                                                 fv = 1.340781e+154                                                 fv = 787.4622                                                 fv = 17.70339                                                 fv = 1.340781e+154                                                 fv = 21.50534                                                 fv = 21.10098                                                 fv = 16.21356                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.56403                                                 fv = 10.3684                                                 fv = 6.375017                                                 fv = 4.272121                                                 fv = 4356674                                                 fv = 1.340781e+154                                                 fv = 1850.28                                                 fv = 93.7074                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.39967                                                 fv = 6.222197                                                 fv = 5.04358                                                 fv = 21.73546                                                 fv = 67.33516                                                 fv = 1624.53                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.38312                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 889.8349                                                 fv = 17.70339                                                 fv = 8.29148                                                 fv = 6.619351                                                 fv = 1.340781e+154                                                 fv = 5.04358                                                 fv = 5.04358                                                 fv = 12.33842                                                 fv = 1.340781e+154                                                 fv = 4.272121                                                 fv = 5.150127                                                 fv = 4.272121                                                 fv = 1.340781e+154                                                 fv = 5.04358                                                 fv = 4.593297                                                 fv = 5.279059                                                 fv = 5.279059                                                 fv = 17.46301                                                 fv = 1.340781e+154                                                 fv = 5.872423                                                 fv = 7939.927                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.272121                                                 fv = 1.340781e+154                                                 fv = 10.56484                                                 fv = 6.430433                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.22245                                                 fv = 1.340781e+154                                                 fv = 5.779022                                                 fv = 29.38384                                                 fv = 1.340781e+154                                                 fv = 9.148105                                                 fv = 4.272121                                                 fv = 4.630574                                                 fv = 4.272121                                                 fv = 4.272121                                                 fv = 4.272121                                                 fv = 4.272121                                                 fv = 1.340781e+154                                                 fv = 43.46716                                                 fv = 20.26166                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.272121                                                 fv = 4.593297                                                 fv = 4.829416                                                 fv = 4.319763                                                 fv = 1.340781e+154                                                 fv = 4.275746                                                 fv = 1.340781e+154                                                 fv = 11.92695                                                 fv = 5.528226                                                 fv = 16.72841                                                 fv = 4.272121                                                 fv = 1.340781e+154                                                 fv = 43.46716 run # 1: f=4.272120787
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.161076                                                 fv = 43.34765                                                 fv = 9.622607                                                 fv = 8.538191                                                 fv = 5.610119                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.2258                                                 fv = 1.340781e+154                                                 fv = 8.526842                                                 fv = 44530.65                                                 fv = 8.736097                                                 fv = 7.307248                                                 fv = 1.340781e+154                                                 fv = 9.862624                                                 fv = 1.340781e+154                                                 fv = 4.067577e+12                                                 fv = 5.610119                                                 fv = 5.610119                                                 fv = 5.610119                                                 fv = 1.340781e+154                                                 fv = 5.754818                                                 fv = 1.340781e+154                                                 fv = 5.754818                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.61326                                                 fv = 336.8699                                                 fv = 20.83078                                                 fv = 25.42241                                                 fv = 5.610119                                                 fv = 19.94322                                                 fv = 5.610119                                                 fv = 6.260379                                                 fv = 23.56049                                                 fv = 5.610119                                                 fv = 1.340781e+154                                                 fv = 6.204866                                                 fv = 5.610119                                                 fv = 5.610119                                                 fv = 23.54094                                                 fv = 14.00662                                                 fv = 141.4458                                                 fv = 1.340781e+154                                                 fv = 89.78746                                                 fv = 25.42241                                                 fv = 545.9753                                                 fv = 20.81963                                                 fv = 1.340781e+154                                                 fv = 8.634672                                                 fv = 8.526842                                                 fv = 15.12169                                                 fv = 17.71352                                                 fv = 10.45891                                                 fv = 12.78056                                                 fv = 10.84506                                                 fv = 12.65033                                                 fv = 5.610119                                                 fv = 5.610119                                                 fv = 5.610119                                                 fv = 7.885248                                                 fv = 5.610119                                                 fv = 11.48472                                                 fv = 5.610119                                                 fv = 5.621053                                                 fv = 5.610119                                                 fv = 10.06212                                                 fv = 14.91479 run # 2: f=5.610119165
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.00306                                                 fv = 43.34765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.00306                                                 fv = 13.35689                                                 fv = 12027.06                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38873565114                                                 fv = 23.81012                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.20578                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24754923                                                 fv = 7097326595                                                 fv = 1.340781e+154                                                 fv = 279250                                                 fv = 151226073                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.35939                                                 fv = 36.74486                                                 fv = 12.8694                                                 fv = 11.43531                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 305910908200                                                 fv = 362890353775                                                 fv = 11.43531                                                 fv = 11.43531                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.43531                                                 fv = 1.340781e+154                                                 fv = 16.77685                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.74486                                                 fv = 12.33856                                                 fv = 23.09044                                                 fv = 11.33341                                                 fv = 9.644044                                                 fv = 9.644044                                                 fv = 1.340781e+154                                                 fv = 38.32466                                                 fv = 36.74486                                                 fv = 2461.99                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9975.56                                                 fv = 9.644044                                                 fv = 9.671226                                                 fv = 9.671226                                                 fv = 1.340781e+154                                                 fv = 687.1748                                                 fv = 16.14367                                                 fv = 8.643068                                                 fv = 8.198407                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.81127                                                 fv = 30956426                                                 fv = 1.340781e+154                                                 fv = 8294383                                                 fv = 1.340781e+154                                                 fv = 34.00109                                                 fv = 8352015                                                 fv = 3573.801                                                 fv = 8.198407                                                 fv = 1.340781e+154                                                 fv = 4641.699                                                 fv = 73.24025                                                 fv = 4641.699                                                 fv = 1.340781e+154                                                 fv = 8.198407                                                 fv = 8.198407                                                 fv = 9.94641                                                 fv = 4641.699                                                 fv = 1.275289e+13                                                 fv = 10.5048                                                 fv = 14.71541                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.74486                                                 fv = 8584325                                                 fv = 1.340781e+154                                                 fv = 8.198407                                                 fv = 9.794044                                                 fv = 8.198407                                                 fv = 18.36449                                                 fv = 1.340781e+154                                                 fv = 104607573                                                 fv = 15.95603                                                 fv = 1468.407                                                 fv = 29.38746                                                 fv = 10.5048                                                 fv = 10.5048                                                 fv = 29.38746                                                 fv = 32.37504                                                 fv = 10.5048                                                 fv = 10.5048                                                 fv = 1.340781e+154                                                 fv = 15.85036                                                 fv = 357.5949                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.35939                                                 fv = 8.198407                                                 fv = 1968.197                                                 fv = 916.2875                                                 fv = 217.1386                                                 fv = 4.467735                                                 fv = 288.7062                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 669.7045                                                 fv = 372.4965                                                 fv = 4.467735                                                 fv = 21.30808                                                 fv = 52.94901                                                 fv = 4.467735                                                 fv = 4.467735                                                 fv = 7.765333                                                 fv = 1.340781e+154                                                 fv = 13.12579                                                 fv = 4.254215                                                 fv = 1.340781e+154                                                 fv = 10.51779                                                 fv = 182569085                                                 fv = 3171.892                                                 fv = 12594.68                                                 fv = 240.4045                                                 fv = 71.61007                                                 fv = 1.340781e+154                                                 fv = 170.271                                                 fv = 4.66549                                                 fv = 611.3037                                                 fv = 5.057802                                                 fv = 1.340781e+154                                                 fv = 4.254215                                                 fv = 8.198407                                                 fv = 29.06549                                                 fv = 67.95999                                                 fv = 67.95999                                                 fv = 12.99423                                                 fv = 1.340781e+154                                                 fv = 9195992265                                                 fv = 1.340781e+154                                                 fv = 13.56608                                                 fv = 4.254215                                                 fv = 273.5577                                                 fv = 8.716828                                                 fv = 4.254215                                                 fv = 10.5048                                                 fv = 21.96969                                                 fv = 8.455833                                                 fv = 10.5048                                                 fv = 10.5048                                                 fv = 8.455833                                                 fv = 1.340781e+154                                                 fv = 10.5048                                                 fv = 5.461606                                                 fv = 1.340781e+154                                                 fv = 396000.3                                                 fv = 13.89792                                                 fv = 8.455833                                                 fv = 8.455833                                                 fv = 4.254215                                                 fv = 4.254215                                                 fv = 54.66989                                                 fv = 28.24284                                                 fv = 4473.474                                                 fv = 4.254215                                                 fv = 1.340781e+154                                                 fv = 10.53178                                                 fv = 4.254215                                                 fv = 4.254215                                                 fv = 5.842282                                                 fv = 13.83068                                                 fv = 4.254215                                                 fv = 8.198407                                                 fv = 4.254215                                                 fv = 88.21216                                                 fv = 214.6924                                                 fv = 29.67354 run # 3: f=4.254214806
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.65847                                                 fv = 43.34765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.07048                                                 fv = 12.99423                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.91479                                                 fv = 1.340781e+154                                                 fv = 9.748006                                                 fv = 13.68513                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.60872                                                 fv = 3221006                                                 fv = 48.09855                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.183746                                                 fv = 102.7972                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.40858                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.5318                                                 fv = 9.654235                                                 fv = 1.340781e+154                                                 fv = 15.5318                                                 fv = 4005973028                                                 fv = 9.183746                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.68877                                                 fv = 23.63714                                                 fv = 9.639113                                                 fv = 1.340781e+154                                                 fv = 9.654235                                                 fv = 30.89858                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 82.15825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.11996                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.077                                                 fv = 22.01137                                                 fv = 7.766573                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.64096                                                 fv = 34.18673                                                 fv = 8125427309                                                 fv = 10.06733                                                 fv = 115.3433                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.42519                                                 fv = 7.766573                                                 fv = 12.29556                                                 fv = 10.26857                                                 fv = 6.870055                                                 fv = 1.340781e+154                                                 fv = 35.84095                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.65471                                                 fv = 6.870055                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 124.9688                                                 fv = 7.973593                                                 fv = 33.99854                                                 fv = 22.01564                                                 fv = 1.340781e+154                                                 fv = 31.85547                                                 fv = 509396308                                                 fv = 11.06097                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.94719                                                 fv = 36.55732                                                 fv = 16.17778                                                 fv = 5.707048                                                 fv = 36.55732                                                 fv = 1.340781e+154                                                 fv = 13.43134                                                 fv = 6.870055                                                 fv = 13.53599                                                 fv = 45.29676                                                 fv = 15.92869                                                 fv = 44.85934                                                 fv = 5.707048                                                 fv = 6.890584                                                 fv = 7.978303                                                 fv = 12.03909                                                 fv = 36.55732                                                 fv = 7.484243                                                 fv = 6.128184                                                 fv = 1.340781e+154                                                 fv = 30.45711                                                 fv = 1.340781e+154                                                 fv = 22.60194                                                 fv = 27.98744                                                 fv = 11.64368                                                 fv = 191.7072                                                 fv = 5.707048                                                 fv = 1.340781e+154                                                 fv = 48.0744                                                 fv = 12.15748                                                 fv = 151.2847                                                 fv = 5.707048                                                 fv = 12.40766                                                 fv = 7.535356                                                 fv = 9.164587                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.707048                                                 fv = 5.707048                                                 fv = 5.707048                                                 fv = 5.707048                                                 fv = 5.707048                                                 fv = 15.74105                                                 fv = 12.73452                                                 fv = 7.885367                                                 fv = 6.966971                                                 fv = 29.68654                                                 fv = 389.0789                                                 fv = 33.58562                                                 fv = 9.621229                                                 fv = 5.707048                                                 fv = 9.884891                                                 fv = 7.048335                                                 fv = 5.393998                                                 fv = 37.22787                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.94782                                                 fv = 4.596765                                                 fv = 9.626684                                                 fv = 2863.48                                                 fv = 3.702044                                                 fv = 1.340781e+154                                                 fv = 3.702044                                                 fv = 4.214034                                                 fv = 3.702044                                                 fv = 3.702044                                                 fv = 3.702044                                                 fv = 6.667295                                                 fv = 6.847464                                                 fv = 4.214801                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.504076                                                 fv = 3.702044                                                 fv = 4.031751                                                 fv = 4.058891                                                 fv = 3.702044                                                 fv = 29.05462                                                 fv = 4.003062                                                 fv = 4.938455                                                 fv = 7.665646                                                 fv = 6.107349                                                 fv = 3.702044                                                 fv = 5.912007                                                 fv = 1.340781e+154                                                 fv = 10.6183                                                 fv = 1.340781e+154                                                 fv = 3.88378                                                 fv = 16.15743 run # 4: f=3.702044101
#> 
#> bb eval : 827
#> best    : 3.702044101
#> worst   : 9.971547804
#> solution: x = ( 0 0 0 0 1 0 0 1 0 1 5 2 1 1 1 10 1 3 1 2 ) f(x) = 3.702044101
#> 
#>                                                 fv = 3.702044                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 43.69288