Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 18.65682                                                 fv = 22.8532                                                 fv = 17.39138                                                 fv = 192426.1                                                 fv = 1.340781e+154                                                 fv = 15.73174                                                 fv = 75.69507                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.17891                                                 fv = 28.99876                                                 fv = 1.340781e+154                                                 fv = 25.16199                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.99876                                                 fv = 20.92039                                                 fv = 17.64372                                                 fv = 20.17891                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 651.0588                                                 fv = 35.97583                                                 fv = 41163912226                                                 fv = 1.340781e+154                                                 fv = 11187416373                                                 fv = 15.73174                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 59.58904                                                 fv = 9304.009                                                 fv = 20.77049                                                 fv = 18.3665                                                 fv = 16.93778                                                 fv = 1.340781e+154                                                 fv = 15.73174                                                 fv = 15.73174                                                 fv = 15.73174                                                 fv = 20.92039                                                 fv = 42.35986                                                 fv = 20.92039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 186.1671                                                 fv = 235.6582                                                 fv = 234.6586                                                 fv = 221286.6                                                 fv = 49293.17                                                 fv = 292470.7                                                 fv = 73.58622                                                 fv = 1.340781e+154                                                 fv = 205.1453                                                 fv = 15.73174                                                 fv = 16.57247                                                 fv = 10.7006                                                 fv = 589598360                                                 fv = 1.340781e+154                                                 fv = 183.4002                                                 fv = 10.75978                                                 fv = 10.7006                                                 fv = 22.73394                                                 fv = 17.83485                                                 fv = 17.66074                                                 fv = 15.38784                                                 fv = 10.7006                                                 fv = 14.22627                                                 fv = 10.7006                                                 fv = 10.7006                                                 fv = 1.340781e+154                                                 fv = 10.7006                                                 fv = 1.340781e+154                                                 fv = 22.77825 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 10.40769                                                 fv = 1.340781e+154                                                 fv = 58.09274                                                 fv = 19.80997 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 10.39259                                                 fv = 1.340781e+154                                                 fv = 324.0635                                                 fv = 33.64392                                                 fv = 24.52354                                                 fv = 5.880786                                                 fv = 1639523                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.25272                                                 fv = 5.880786                                                 fv = 5.880786                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.992523                                                 fv = 12.80292                                                 fv = 20.00785                                                 fv = 7.800444                                                 fv = 19.32054                                                 fv = 19.32054                                                 fv = 25.71346                                                 fv = 8.595962                                                 fv = 16.43793                                                 fv = 16.43793                                                 fv = 59.85803                                                 fv = 35822992                                                 fv = 17.90456                                                 fv = 12.81304                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.38847                                                 fv = 6.423579                                                 fv = 6.584983                                                 fv = 8.991455                                                 fv = 809.7649                                                 fv = 11.14438 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 11.99142                                                 fv = 14.06234                                                 fv = 10.3372                                                 fv = 5.880786                                                 fv = 22.46281                                                 fv = 1.340781e+154                                                 fv = 5.880786                                                 fv = 5.880786                                                 fv = 5.880786                                                 fv = 1.340781e+154                                                 fv = 14.70649                                                 fv = 14.70649                                                 fv = 13.13497                                                 fv = 13.25688                                                 fv = 15.20384                                                 fv = 4453.799                                                 fv = 7.305675                                                 fv = 12.59077                                                 fv = 10.493 run # 0: f=5.880786213
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.05669                                                 fv = 33.4571                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.250985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.22264                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.85413                                                 fv = 4452718                                                 fv = 3.587011e+13                                                 fv = 9.029137                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.5566                                                 fv = 1.340781e+154                                                 fv = 11.209                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 971087.7                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 8.337506                                                 fv = 1.340781e+154                                                 fv = 23.77511                                                 fv = 23.77511                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 1638.529                                                 fv = 1.340781e+154                                                 fv = 11.209                                                 fv = 204.0034                                                 fv = 19898.53                                                 fv = 6.757251                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 28.24042                                                 fv = 36.38196                                                 fv = 9.16458                                                 fv = 1.340781e+154                                                 fv = 7.558134                                                 fv = 6.327996                                                 fv = 6.327996                                                 fv = 10.22746                                                 fv = 20.12833                                                 fv = 27.28357                                                 fv = 13.26575                                                 fv = 17.67795                                                 fv = 449.2781                                                 fv = 19.0728                                                 fv = 9.075095                                                 fv = 7.1293                                                 fv = 43.82729                                                 fv = 6.327996                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 13.35814                                                 fv = 1.340781e+154                                                 fv = 6.327996                                                 fv = 15.41744                                                 fv = 6.327996                                                 fv = 6.327996                                                 fv = 6.327996                                                 fv = 11.26356                                                 fv = 1.340781e+154                                                 fv = 23.77511                                                 fv = 9.16458                                                 fv = 16.35746                                                 fv = 5.038619                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.038619                                                 fv = 9.167606                                                 fv = 4.118503                                                 fv = 96.94691                                                 fv = 1.340781e+154                                                 fv = 27.32583                                                 fv = 4.118503                                                 fv = 7.77645                                                 fv = 7.146508                                                 fv = 4.118503                                                 fv = 4.118503                                                 fv = 3.105923                                                 fv = 50.29857                                                 fv = 1.340781e+154                                                 fv = 23.77511                                                 fv = 4.118503                                                 fv = 5.150225                                                 fv = 3.105923                                                 fv = 3.105923                                                 fv = 3.105923                                                 fv = 3.105923                                                 fv = 1.340781e+154                                                 fv = 3.917205                                                 fv = 10.93183                                                 fv = 1.340781e+154                                                 fv = 3.105923                                                 fv = 9.265621                                                 fv = 3.105923                                                 fv = 5.25699                                                 fv = 3.881901                                                 fv = 9.748432                                                 fv = 4.767271                                                 fv = 8.319443                                                 fv = 705.9518                                                 fv = 7.398168                                                 fv = 3.849495                                                 fv = 5.889869                                                 fv = 6.011157                                                 fv = 5.868457                                                 fv = 61.05796                                                 fv = 9.99838                                                 fv = 34.2466 run # 2: f=3.105923386
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10289169                                                 fv = 126090454                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10289169                                                 fv = 119157196                                                 fv = 6783581                                                 fv = 1100329005                                                 fv = 772601567                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6783581                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.471693e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 921880119                                                 fv = 1.340781e+154                                                 fv = 268715.1                                                 fv = 23.12663                                                 fv = 28.26896                                                 fv = 12.09426                                                 fv = 94.7552                                                 fv = 12906.84                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28321393704                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.37289                                                 fv = 16.31734                                                 fv = 10.03592                                                 fv = 10.45319                                                 fv = 13.90315                                                 fv = 3847950040                                                 fv = 1.340781e+154                                                 fv = 1099998636                                                 fv = 744294355                                                 fv = 104.8673                                                 fv = 20.63587                                                 fv = 9298484                                                 fv = 22489.87                                                 fv = 1.340781e+154                                                 fv = 13.88727                                                 fv = 1.340781e+154                                                 fv = 10.04815                                                 fv = 123160723504                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45072830                                                 fv = 104.8673                                                 fv = 1.340781e+154                                                 fv = 38656.23                                                 fv = 9298484                                                 fv = 1.340781e+154                                                 fv = 60666.56                                                 fv = 10.04815                                                 fv = 14.87105                                                 fv = 1.340781e+154                                                 fv = 10.04815                                                 fv = 11.02721                                                 fv = 1.340781e+154                                                 fv = 33.08132                                                 fv = 77170977547                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 585342.4                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.45319                                                 fv = 33.4571                                                 fv = 175449                                                 fv = 9.464599                                                 fv = 264058.7                                                 fv = 1.340781e+154                                                 fv = 17.36695                                                 fv = 90793257                                                 fv = 10.45319                                                 fv = 9.464599                                                 fv = 72.23731                                                 fv = 1.340781e+154                                                 fv = 9.464599                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.03592                                                 fv = 9.464599                                                 fv = 10.03592                                                 fv = 10.03592                                                 fv = 653901                                                 fv = 10.59414                                                 fv = 9.464599                                                 fv = 9.464599                                                 fv = 9.393303                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.464599                                                 fv = 1.340781e+154                                                 fv = 9.393303                                                 fv = 9.393303                                                 fv = 9.393303                                                 fv = 9.393303                                                 fv = 9.393303                                                 fv = 13.7968                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.393303                                                 fv = 17.40245                                                 fv = 9.393303                                                 fv = 1.340781e+154                                                 fv = 33.51639                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 17.36695                                                 fv = 19557.97                                                 fv = 21.38182                                                 fv = 21.49572                                                 fv = 9.334355                                                 fv = 10.0614                                                 fv = 1.340781e+154                                                 fv = 9.688863                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 52.00923                                                 fv = 14.16147                                                 fv = 12.45538                                                 fv = 12.95978                                                 fv = 83746.25                                                 fv = 14.16147                                                 fv = 10.6293                                                 fv = 24                                                 fv = 27.38961                                                 fv = 117645337                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.36695                                                 fv = 1.340781e+154                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 9.890605                                                 fv = 9.334355                                                 fv = 9.334355                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 9.334355                                                 fv = 10.13741                                                 fv = 12.41652                                                 fv = 9.334355                                                 fv = 9.725483                                                 fv = 9.334355                                                 fv = 14.16147                                                 fv = 14.16147                                                 fv = 21.49572                                                 fv = 35.09674                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 30.94083                                                 fv = 13.44688                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.334355                                                 fv = 9.334355                                                 fv = 9.393303                                                 fv = 9.334355                                                 fv = 10.13741                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.73187                                                 fv = 375.602                                                 fv = 9.725483                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.05762                                                 fv = 9.334355                                                 fv = 9.334355                                                 fv = 9.334355                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.890605                                                 fv = 14.16147                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 373.3308                                                 fv = 11.73412 run # 4: f=9.334354982
#> 
#> bb eval : 580
#> best    : 3.105923386
#> worst   : 1.340780793e+154
#> solution: x = ( 0 4 0 3 0 1 1 0 0 0 3 1 1 4 1 1 1 1 1 3 ) f(x) = 3.105923386
#> 
#>                                                 fv = 3.105923                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 28.39625