Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 40.29334                                                 fv = 28.7804                                                 fv = 1.340781e+154                                                 fv = 48.5408                                                 fv = 1.340781e+154                                                 fv = 166.2542                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5134.761                                                 fv = 28.7804                                                 fv = 1.340781e+154                                                 fv = 8.381291                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10272.41                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.381291                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.989643                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20593.05                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20593.05                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.220899                                                 fv = 29.14265                                                 fv = 1.340781e+154                                                 fv = 245.2897                                                 fv = 1.340781e+154                                                 fv = 7.463571                                                 fv = 85.57507                                                 fv = 1049.909                                                 fv = 17389.59                                                 fv = 9.531556                                                 fv = 1.340781e+154                                                 fv = 7.685514                                                 fv = 5.44475                                                 fv = 5.839949                                                 fv = 10.50825                                                 fv = 4.189324                                                 fv = 4.189324                                                 fv = 1.340781e+154                                                 fv = 17.27696                                                 fv = 1.340781e+154                                                 fv = 130586513704                                                 fv = 1521184                                                 fv = 353785.9                                                 fv = 855677909                                                 fv = 14.17659                                                 fv = 14.17659                                                 fv = 14.17659                                                 fv = 1.340781e+154                                                 fv = 93.9884                                                 fv = 8.030993                                                 fv = 14.17659                                                 fv = 27.67323                                                 fv = 6.010594                                                 fv = 130313.9                                                 fv = 5.220899                                                 fv = 6.165176                                                 fv = 5.44475                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.41761                                                 fv = 4.148452                                                 fv = 4.148452                                                 fv = 1.340781e+154                                                 fv = 16.29829                                                 fv = 4.148452                                                 fv = 4.148452                                                 fv = 3.836237                                                 fv = 3.836237                                                 fv = 1.340781e+154                                                 fv = 16.29829                                                 fv = 4.156702                                                 fv = 4.156702                                                 fv = 5.097341                                                 fv = 14.83721                                                 fv = 13.78177                                                 fv = 4.156702                                                 fv = 7.476284                                                 fv = 12.02288                                                 fv = 3.836237                                                 fv = 4.799626                                                 fv = 1871.718                                                 fv = 4.079245                                                 fv = 3.836237                                                 fv = 4.079245                                                 fv = 4.079245                                                 fv = 7.324411                                                 fv = 7.151753                                                 fv = 4.079245                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.700221                                                 fv = 65.64681                                                 fv = 4.495207                                                 fv = 1.340781e+154                                                 fv = 3.836237                                                 fv = 3.836237                                                 fv = 3.836237                                                 fv = 3.836237                                                 fv = 3.836237                                                 fv = 4.079245                                                 fv = 3.407538                                                 fv = 1.340781e+154                                                 fv = 8.700221                                                 fv = 4.791271                                                 fv = 3.407538                                                 fv = 1.340781e+154                                                 fv = 3.407538                                                 fv = 3.407538                                                 fv = 3.407538                                                 fv = 7.275214                                                 fv = 4.222651                                                 fv = 3.407538                                                 fv = 4.055134                                                 fv = 3.432489                                                 fv = 4.055134                                                 fv = 5.259418                                                 fv = 7.560528                                                 fv = 13.13568                                                 fv = 1.340781e+154                                                 fv = 3.439557                                                 fv = 153.0892                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.58752                                                 fv = 12.28772                                                 fv = 3.302857                                                 fv = 14.73176                                                 fv = 1.340781e+154                                                 fv = 68.28408                                                 fv = 3.302857                                                 fv = 4.998616                                                 fv = 3.302857                                                 fv = 3.302857                                                 fv = 4.372075                                                 fv = 3.302857                                                 fv = 3.302857                                                 fv = 1.340781e+154                                                 fv = 3.302857                                                 fv = 1.340781e+154                                                 fv = 3.58709                                                 fv = 3.302857                                                 fv = 3.302857                                                 fv = 3.302857                                                 fv = 5.672259                                                 fv = 4.373204                                                 fv = 4.176361                                                 fv = 10.34038                                                 fv = 10.34038                                                 fv = 17.84704                                                 fv = 6.709526                                                 fv = 22.91489                                                 fv = 7.263992                                                 fv = 1.340781e+154                                                 fv = 13.67024                                                 fv = 67.94281                                                 fv = 63.89655                                                 fv = 1.340781e+154                                                 fv = 3.302857                                                 fv = 4.176361                                                 fv = 3.302857                                                 fv = 3.407538                                                 fv = 3.302857                                                 fv = 3.302857                                                 fv = 3.641615                                                 fv = 6.763116                                                 fv = 10.34038                                                 fv = 41.95737                                                 fv = 195.1968 run # 0: f=3.302856713
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.724392                                                 fv = 29.14265                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.011474                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.90947                                                 fv = 1.340781e+154                                                 fv = 21016.23                                                 fv = 561729731                                                 fv = 1.340781e+154                                                 fv = 30904102155                                                 fv = 192.5094                                                 fv = 5.561487                                                 fv = 406289354                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.87398                                                 fv = 14.49532                                                 fv = 25.42263                                                 fv = 1.340781e+154                                                 fv = 713.6872 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 22.62226 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 22.62226 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 22.62226                                                 fv = 1.340781e+154                                                 fv = 27.19272                                                 fv = 7.438801                                                 fv = 267.5184                                                 fv = 34527.07                                                 fv = 8.085778                                                 fv = 446066.2                                                 fv = 9.914261                                                 fv = 5.561487                                                 fv = 13.6424                                                 fv = 5.011474                                                 fv = 5.011474                                                 fv = 5.011474                                                 fv = 6.702231                                                 fv = 5.011474                                                 fv = 1.340781e+154                                                 fv = 5.011474                                                 fv = 642.0605                                                 fv = 15.07743                                                 fv = 1.340781e+154                                                 fv = 17.30075                                                 fv = 1.340781e+154                                                 fv = 29.14265                                                 fv = 12.14299                                                 fv = 5.330708                                                 fv = 55.11459                                                 fv = 7.713087                                                 fv = 5.56804                                                 fv = 524.1442                                                 fv = 5.822739                                                 fv = 6.388105                                                 fv = 7.464863                                                 fv = 5.011474                                                 fv = 5.011474                                                 fv = 11.31352                                                 fv = 5.011474                                                 fv = 50.52319                                                 fv = 5.011474                                                 fv = 5.011474                                                 fv = 1.340781e+154                                                 fv = 8.055167                                                 fv = 10.26378 run # 1: f=5.011474109
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.45295                                                 fv = 29.14265                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.45295                                                 fv = 62.53812                                                 fv = 34.49744                                                 fv = 386.5138                                                 fv = 803.5538                                                 fv = 353.6798                                                 fv = 66285.01                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.15608                                                 fv = 1.340781e+154                                                 fv = 30.58023                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42.36128                                                 fv = 1.340781e+154                                                 fv = 10.11961                                                 fv = 1295.005                                                 fv = 4.658556                                                 fv = 13.41914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.15548                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.82799                                                 fv = 12.90947                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.90947                                                 fv = 12.90947                                                 fv = 1.340781e+154                                                 fv = 540904.3                                                 fv = 10.99791                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.14265                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.24603                                                 fv = 4.658556                                                 fv = 26.03207                                                 fv = 35.27327 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 6.242542                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.592377                                                 fv = 1.340781e+154                                                 fv = 2.714417e+12                                                 fv = 1.340781e+154                                                 fv = 4.658556                                                 fv = 1.340781e+154                                                 fv = 7.993087                                                 fv = 1312.124                                                 fv = 23.98006                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.07668                                                 fv = 9.50101                                                 fv = 1.340781e+154                                                 fv = 10.82983                                                 fv = 1929.495                                                 fv = 6.714956                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.658556                                                 fv = 6.880467                                                 fv = 1587.49                                                 fv = 21.622                                                 fv = 9.60908                                                 fv = 1.340781e+154                                                 fv = 5.381845                                                 fv = 4.658556                                                 fv = 128.5408                                                 fv = 1.340781e+154                                                 fv = 1388007                                                 fv = 121765.6                                                 fv = 11.59329                                                 fv = 1.340781e+154                                                 fv = 5.291765                                                 fv = 4.658556                                                 fv = 8.464942                                                 fv = 5.205252                                                 fv = 64.44292 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 7.461854                                                 fv = 1.340781e+154                                                 fv = 9.160881                                                 fv = 6.733251                                                 fv = 176.7507 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 5.653846                                                 fv = 1.340781e+154                                                 fv = 6.002717                                                 fv = 4.658556                                                 fv = 10.38286                                                 fv = 1.340781e+154                                                 fv = 7.302816                                                 fv = 4.658556                                                 fv = 6.242542                                                 fv = 4.658556                                                 fv = 14.55838 run # 2: f=4.658555599
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.72486                                                 fv = 29.14265                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.14265                                                 fv = 7.278436                                                 fv = 15.5922                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.809229                                                 fv = 310.1411                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.809229                                                 fv = 1.340781e+154                                                 fv = 12.68003                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.809229                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.599443                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 9.691331                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.14299                                                 fv = 7.278436                                                 fv = 33.28373                                                 fv = 9.1993                                                 fv = 9.1993                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.83022                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.78928                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 11.50595                                                 fv = 7.278436                                                 fv = 1.340781e+154                                                 fv = 8.522928                                                 fv = 9.018991                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.14299                                                 fv = 8.730183                                                 fv = 9.639033                                                 fv = 9.757947                                                 fv = 8.882761                                                 fv = 1.340781e+154                                                 fv = 8.833996                                                 fv = 8.036273                                                 fv = 28.39887                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 1.340781e+154                                                 fv = 8.522928                                                 fv = 8.182121                                                 fv = 8.522928                                                 fv = 9.36795                                                 fv = 10.97417                                                 fv = 12.89338                                                 fv = 15.6304                                                 fv = 1.340781e+154                                                 fv = 8.099008                                                 fv = 8.099008                                                 fv = 8.45052                                                 fv = 9.841897                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 9.652073                                                 fv = 7.889757                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 7.278436                                                 fv = 10.3262                                                 fv = 8.713839                                                 fv = 7.809229                                                 fv = 8.141348                                                 fv = 9.1993                                                 fv = 8.429536                                                 fv = 8.034249                                                 fv = 8.910415 run # 4: f=7.278436057
#> 
#> bb eval : 606
#> best    : 3.302856713
#> worst   : 1.340780793e+154
#> solution: x = ( 1 0 2 4 0 0 0 0 1 0 1 1 1 1 1 1 2 8 1 5 ) f(x) = 3.302856713
#> 
#>                                                 fv = 3.302857                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.97207