Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 26.72842                                                 fv = 12.25684                                                 fv = 1.340781e+154                                                 fv = 27.02031                                                 fv = 1.340781e+154                                                 fv = 34.63835                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 67.28523                                                 fv = 12.25684                                                 fv = 1.340781e+154                                                 fv = 1157.469                                                 fv = 34.63835                                                 fv = 1.340781e+154                                                 fv = 44.32702                                                 fv = 44.32702                                                 fv = 426.313                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.702617                                                 fv = 13.95445                                                 fv = 1.340781e+154                                                 fv = 8.169511                                                 fv = 432.3                                                 fv = 12.96643                                                 fv = 11.08856                                                 fv = 11281571325                                                 fv = 8722889894                                                 fv = 1.340781e+154                                                 fv = 6.91849                                                 fv = 1.340781e+154                                                 fv = 1.796108e+12                                                 fv = 6.91849                                                 fv = 6.91849                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.91849                                                 fv = 7.696196                                                 fv = 6.629325                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.19261                                                 fv = 7.737218                                                 fv = 6.601346                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.77184                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.75682                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.300307                                                 fv = 6.522333                                                 fv = 1.340781e+154                                                 fv = 8.059334                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.494871                                                 fv = 7.183312                                                 fv = 7.330447                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.25107                                                 fv = 10.03516                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7183.19                                                 fv = 1.340781e+154                                                 fv = 8.059334                                                 fv = 516.9614                                                 fv = 818.6586                                                 fv = 8.080245                                                 fv = 1.340781e+154                                                 fv = 11.65209                                                 fv = 8.236435                                                 fv = 1.340781e+154                                                 fv = 6.300307                                                 fv = 6.522333                                                 fv = 6.131592                                                 fv = 1.340781e+154                                                 fv = 8.247437                                                 fv = 6.131592                                                 fv = 6.131592                                                 fv = 7.131249                                                 fv = 1967.793                                                 fv = 1.340781e+154                                                 fv = 10.02127                                                 fv = 1.340781e+154                                                 fv = 11.28009                                                 fv = 5.949707                                                 fv = 7.595482                                                 fv = 1.340781e+154                                                 fv = 8.247437                                                 fv = 9029.366                                                 fv = 5.949707                                                 fv = 7.595482                                                 fv = 11.38642                                                 fv = 7.761382                                                 fv = 6.729638                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.949707                                                 fv = 6.819823                                                 fv = 13.32214                                                 fv = 1.340781e+154                                                 fv = 5.949707                                                 fv = 1.340781e+154                                                 fv = 14.31244                                                 fv = 11.87204                                                 fv = 1.340781e+154                                                 fv = 8.896071                                                 fv = 746.253                                                 fv = 5.949707                                                 fv = 72560.47                                                 fv = 1.340781e+154                                                 fv = 7.916997                                                 fv = 5.949707                                                 fv = 11.73705                                                 fv = 7.839873                                                 fv = 1.340781e+154                                                 fv = 7.380312                                                 fv = 1.340781e+154                                                 fv = 8.787365                                                 fv = 8.579119                                                 fv = 8.146493                                                 fv = 6.021105                                                 fv = 9.11811                                                 fv = 6.142533                                                 fv = 1.340781e+154                                                 fv = 6.131592                                                 fv = 11.51007                                                 fv = 5.949707                                                 fv = 5.949707                                                 fv = 11.51007                                                 fv = 5.949707                                                 fv = 6.131592                                                 fv = 9.021052                                                 fv = 1.340781e+154                                                 fv = 6.300307                                                 fv = 5.949707                                                 fv = 1.340781e+154                                                 fv = 7.595482                                                 fv = 7.965022                                                 fv = 6.300307                                                 fv = 6.400671                                                 fv = 6.726472                                                 fv = 6.881773                                                 fv = 5.949707                                                 fv = 5.949707                                                 fv = 5.949707                                                 fv = 6.643488                                                 fv = 1.340781e+154                                                 fv = 7.209735                                                 fv = 6.131592                                                 fv = 11.73916 run # 0: f=5.949706698
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.737437                                                 fv = 37.87295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.35616                                                 fv = 7.697961                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.697961                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.64577                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.03987                                                 fv = 1.340781e+154                                                 fv = 137934270557                                                 fv = 7.697961                                                 fv = 7.697961                                                 fv = 49.60774                                                 fv = 12.66047                                                 fv = 1.340781e+154                                                 fv = 20.24331                                                 fv = 20.24331                                                 fv = 32.80419                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.108499                                                 fv = 10.19163                                                 fv = 10.6152                                                 fv = 85.55636                                                 fv = 33.41663                                                 fv = 66.09475                                                 fv = 7.697961                                                 fv = 15063.18                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.45034                                                 fv = 2705.723                                                 fv = 37.87295                                                 fv = 32.80419                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.858337                                                 fv = 11.08929                                                 fv = 1.340781e+154                                                 fv = 24.05282                                                 fv = 11.14258                                                 fv = 9.581272                                                 fv = 7.697961                                                 fv = 7.697961                                                 fv = 7.697961                                                 fv = 7.697961                                                 fv = 1.340781e+154                                                 fv = 8.569362                                                 fv = 7.697961                                                 fv = 6.173398                                                 fv = 6352580                                                 fv = 1.340781e+154                                                 fv = 36.60444                                                 fv = 36.60444                                                 fv = 8.865102                                                 fv = 1.340781e+154                                                 fv = 41.28117                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 1.340781e+154                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 8.712489                                                 fv = 7.936711                                                 fv = 11.30047                                                 fv = 11.30047                                                 fv = 10.80734                                                 fv = 14.94852                                                 fv = 11.30047                                                 fv = 9.196445                                                 fv = 9.196445                                                 fv = 34.35616                                                 fv = 8605897                                                 fv = 163.9443                                                 fv = 1.340781e+154                                                 fv = 157.1151                                                 fv = 7.034743                                                 fv = 11.40621                                                 fv = 14.81832                                                 fv = 6.310384                                                 fv = 12.00009                                                 fv = 7.785883                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 6.901795                                                 fv = 1.340781e+154                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 6.173398                                                 fv = 9.672075 run # 1: f=6.173397739
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.95201                                                 fv = 37.87295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.62643                                                 fv = 5.95201                                                 fv = 7.438213                                                 fv = 1.340781e+154                                                 fv = 7.577606                                                 fv = 9.137548                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 352943059946                                                 fv = 13.82927                                                 fv = 374605919137                                                 fv = 1.340781e+154                                                 fv = 12.91331                                                 fv = 18.24465                                                 fv = 5.719613                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 32.80419                                                 fv = 5.901514                                                 fv = 6.524983                                                 fv = 8.438987                                                 fv = 11.00634                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.616853                                                 fv = 37481.89                                                 fv = 3614177503                                                 fv = 305800845523                                                 fv = 1.340781e+154                                                 fv = 1007976110                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 7932232                                                 fv = 6.524983                                                 fv = 5.616853                                                 fv = 86.44213                                                 fv = 1.340781e+154                                                 fv = 13.05615                                                 fv = 276.8809                                                 fv = 141640.8                                                 fv = 1.340781e+154                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 1114.683                                                 fv = 5.719613                                                 fv = 443.4356                                                 fv = 5.719613                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 107.9963                                                 fv = 1.340781e+154                                                 fv = 12.97326                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 12.77529                                                 fv = 12.70719                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 7.177947                                                 fv = 6.931262                                                 fv = 8.7677                                                 fv = 7.352464                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 6.010371                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 6.239156                                                 fv = 5.616853                                                 fv = 5.95201                                                 fv = 45.96374                                                 fv = 7.398619                                                 fv = 5.616853                                                 fv = 8.661063                                                 fv = 8.593387                                                 fv = 7.221679                                                 fv = 5.616853                                                 fv = 6.145414                                                 fv = 39.02157                                                 fv = 7.405823                                                 fv = 7.726913                                                 fv = 305447.1                                                 fv = 1.340781e+154                                                 fv = 5.616853                                                 fv = 6.752115                                                 fv = 1.340781e+154                                                 fv = 5.901514                                                 fv = 5.96391                                                 fv = 6.967179                                                 fv = 6.410097                                                 fv = 5.616853                                                 fv = 1.340781e+154                                                 fv = 6.642782                                                 fv = 6.524983                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 5.616853                                                 fv = 6.750164                                                 fv = 10.97478 run # 2: f=5.616852518
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 78.99423                                                 fv = 37.87295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.35616                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.31509                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.529961e+13                                                 fv = 15.25276                                                 fv = 104.7544                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 889799805                                                 fv = 12.04904                                                 fv = 1459472                                                 fv = 2513.607                                                 fv = 48660857                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 204.6001                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.21407                                                 fv = 94684383                                                 fv = 14.32873                                                 fv = 1.340781e+154                                                 fv = 24.88942                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.16929                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.04904                                                 fv = 1.340781e+154                                                 fv = 11.29828                                                 fv = 27.48067                                                 fv = 40.76063                                                 fv = 1.340781e+154                                                 fv = 535007731                                                 fv = 1.340781e+154                                                 fv = 11.29828                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.29828                                                 fv = 1.340781e+154                                                 fv = 8.597908                                                 fv = 8.597908                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.50179                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.83502                                                 fv = 34.35616                                                 fv = 1.340781e+154                                                 fv = 140.9892                                                 fv = 12.94458                                                 fv = 7.238673                                                 fv = 1.340781e+154                                                 fv = 9.032058                                                 fv = 7.238673                                                 fv = 7.643511                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 125.3519                                                 fv = 6.659756                                                 fv = 1.173905e+13                                                 fv = 1.340781e+154                                                 fv = 39.24492                                                 fv = 470.2753                                                 fv = 1.340781e+154                                                 fv = 6.659756                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.06813                                                 fv = 1.340781e+154                                                 fv = 19.18838                                                 fv = 31.72816                                                 fv = 1.340781e+154                                                 fv = 16.43952                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.17465                                                 fv = 28.17465                                                 fv = 44.93533                                                 fv = 7.270145                                                 fv = 6.659756                                                 fv = 1.340781e+154                                                 fv = 23.35572                                                 fv = 35.33825                                                 fv = 27.22399                                                 fv = 6.659756                                                 fv = 12.15687                                                 fv = 1.340781e+154                                                 fv = 54.65523                                                 fv = 9.34732                                                 fv = 46.62047                                                 fv = 16.96554                                                 fv = 65.07652                                                 fv = 1.340781e+154                                                 fv = 305387.8                                                 fv = 205258555                                                 fv = 18208.31                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.41262                                                 fv = 16.63058                                                 fv = 303.6718                                                 fv = 83469.71                                                 fv = 6.659756                                                 fv = 39.27751                                                 fv = 792.1719                                                 fv = 295.7817                                                 fv = 72.04511                                                 fv = 5962.592                                                 fv = 79.14535                                                 fv = 6.659756                                                 fv = 1.340781e+154                                                 fv = 6.659756                                                 fv = 39.19015                                                 fv = 1.340781e+154                                                 fv = 17.89489                                                 fv = 1.340781e+154                                                 fv = 14.29281                                                 fv = 24421741                                                 fv = 12.36201                                                 fv = 1.340781e+154                                                 fv = 80.11631                                                 fv = 11.40968                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 10.10942                                                 fv = 1.340781e+154                                                 fv = 27.98013                                                 fv = 118.8996                                                 fv = 47.85501                                                 fv = 35.76031                                                 fv = 71.97603                                                 fv = 1652.298                                                 fv = 6.659756                                                 fv = 6.256073                                                 fv = 2311.544                                                 fv = 1.340781e+154                                                 fv = 22.58335                                                 fv = 6.256073                                                 fv = 5.079932                                                 fv = 4.721331                                                 fv = 1.340781e+154                                                 fv = 7.503974                                                 fv = 1.340781e+154                                                 fv = 3.969078                                                 fv = 417.9764                                                 fv = 1.340781e+154                                                 fv = 8.551054                                                 fv = 5.885283                                                 fv = 4.83679                                                 fv = 15.37307                                                 fv = 11.56385                                                 fv = 1.340781e+154                                                 fv = 3.969078                                                 fv = 3.969078                                                 fv = 3.969078                                                 fv = 7.274634                                                 fv = 24.82188                                                 fv = 26.00519                                                 fv = 17.69003                                                 fv = 22633.78                                                 fv = 1479.895                                                 fv = 5.429966                                                 fv = 23.05801                                                 fv = 1437.638                                                 fv = 124446                                                 fv = 1.340781e+154                                                 fv = 2070.004                                                 fv = 33.67333                                                 fv = 18.60347                                                 fv = 3.969078                                                 fv = 5.147238                                                 fv = 3.969078                                                 fv = 3.045069                                                 fv = 4.419081                                                 fv = 1.340781e+154                                                 fv = 3755.08                                                 fv = 14.73649                                                 fv = 3.045069                                                 fv = 3.045069                                                 fv = 3.045069                                                 fv = 3.045069                                                 fv = 3.778183                                                 fv = 3.969078                                                 fv = 26.78821                                                 fv = 4.106132                                                 fv = 3.363109                                                 fv = 4.419081                                                 fv = 5.192558                                                 fv = 4.204468                                                 fv = 9.11986                                                 fv = 18.21505                                                 fv = 4.476139                                                 fv = 1471.448                                                 fv = 7.934788 run # 4: f=3.045069052
#> 
#> bb eval : 741
#> best    : 3.045069052
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 1 0 1 0 1 1 0 2 1 2 1 4 1 1 3 5 1 1 ) f(x) = 3.045069052
#> 
#>                                                 fv = 3.045069                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...          

# Score the predictions
predictions$score()
#> regr.mse 
#>  6.96899