Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 17.13948                                                 fv = 29.70809                                                 fv = 15.98118                                                 fv = 192552511                                                 fv = 1.340781e+154                                                 fv = 32.1428                                                 fv = 101.7207                                                 fv = 3802540557                                                 fv = 1.340781e+154                                                 fv = 801.8756                                                 fv = 44.95788                                                 fv = 1.340781e+154                                                 fv = 2701.771                                                 fv = 1.340781e+154                                                 fv = 1096.069                                                 fv = 1.340781e+154                                                 fv = 101.7207                                                 fv = 101.7207                                                 fv = 1.340781e+154                                                 fv = 9614.76                                                 fv = 1.340781e+154                                                 fv = 45.49743                                                 fv = 2909310863                                                 fv = 20714.52                                                 fv = 52.95591                                                 fv = 21.13255                                                 fv = 15.98118                                                 fv = 41.75259                                                 fv = 52.95591                                                 fv = 707.51                                                 fv = 205.7006                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.98118                                                 fv = 299.4058                                                 fv = 218874.8                                                 fv = 62.19576                                                 fv = 12.51649                                                 fv = 18389.95                                                 fv = 1.340781e+154                                                 fv = 19265560                                                 fv = 56.16273                                                 fv = 1.340781e+154                                                 fv = 17.73976                                                 fv = 15.1394                                                 fv = 390.1962                                                 fv = 5661.094                                                 fv = 1.340781e+154                                                 fv = 17.73976                                                 fv = 15.87003                                                 fv = 17.73976                                                 fv = 31.15423                                                 fv = 28.47619                                                 fv = 80.57069                                                 fv = 12.51649                                                 fv = 194.2618                                                 fv = 390.1962                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4192.693                                                 fv = 12.51649                                                 fv = 1.340781e+154                                                 fv = 521250.1                                                 fv = 17.73976                                                 fv = 23.71542 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 25.37347                                                 fv = 8.264869                                                 fv = 11.82194                                                 fv = 1.340781e+154                                                 fv = 63.34381                                                 fv = 11.08272                                                 fv = 8.09306                                                 fv = 1.340781e+154                                                 fv = 6.542732                                                 fv = 1.340781e+154                                                 fv = 10.36254                                                 fv = 14.14555                                                 fv = 6.542732                                                 fv = 1.340781e+154                                                 fv = 10.32535                                                 fv = 8.793128                                                 fv = 8.60341                                                 fv = 6.542732                                                 fv = 10.93094                                                 fv = 6.542732                                                 fv = 8.211189                                                 fv = 6.542732                                                 fv = 8.234304                                                 fv = 6.725125                                                 fv = 6.767635                                                 fv = 5.474821                                                 fv = 1.340781e+154                                                 fv = 11.08972                                                 fv = 76.13888                                                 fv = 5.474821                                                 fv = 5.474821                                                 fv = 13.4303                                                 fv = 6.542732                                                 fv = 1.340781e+154                                                 fv = 5.474821                                                 fv = 6.193338                                                 fv = 8.104457                                                 fv = 16.00313                                                 fv = 30.93474                                                 fv = 6.193338                                                 fv = 19.98063                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.474821                                                 fv = 5.763323                                                 fv = 5.687994                                                 fv = 5.64246                                                 fv = 5.474821                                                 fv = 5.474821                                                 fv = 5.474821                                                 fv = 6.542732                                                 fv = 6.193338                                                 fv = 1.340781e+154                                                 fv = 4.258467                                                 fv = 4118.483                                                 fv = 1.340781e+154                                                 fv = 5.966694                                                 fv = 4.983613                                                 fv = 8.272844                                                 fv = 4.258467                                                 fv = 6.693879                                                 fv = 1.340781e+154                                                 fv = 23.72031                                                 fv = 4.74202                                                 fv = 8.136521                                                 fv = 4.258467                                                 fv = 1.340781e+154                                                 fv = 4.258467                                                 fv = 18.22202                                                 fv = 23.72031                                                 fv = 7.668986                                                 fv = 6.225083                                                 fv = 6.468373                                                 fv = 7.088825                                                 fv = 9.759152                                                 fv = 127.8893                                                 fv = 28.60557                                                 fv = 8.914819                                                 fv = 7.730026                                                 fv = 16.41052                                                 fv = 18.22202                                                 fv = 15.86111                                                 fv = 4.89934                                                 fv = 155052374802                                                 fv = 1.340781e+154                                                 fv = 4.511777                                                 fv = 15.20634                                                 fv = 1.340781e+154                                                 fv = 4.258467                                                 fv = 7.643994                                                 fv = 7.726918                                                 fv = 1.340781e+154                                                 fv = 4.258467                                                 fv = 5.691522                                                 fv = 1.340781e+154                                                 fv = 4.258467                                                 fv = 18.22202                                                 fv = 6.422061                                                 fv = 6.780682                                                 fv = 28.60557                                                 fv = 116860.7                                                 fv = 1129.244                                                 fv = 1.340781e+154                                                 fv = 4.788209                                                 fv = 4.511777                                                 fv = 8.495052                                                 fv = 4.557084                                                 fv = 5.691522                                                 fv = 21.09076                                                 fv = 4.258467                                                 fv = 9.230498 run # 0: f=4.258467182
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.906197                                                 fv = 32.95835                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.906197                                                 fv = 1.340781e+154                                                 fv = 25.12582                                                 fv = 4.506868                                                 fv = 132.5612                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7449519                                                 fv = 1062813712                                                 fv = 1.340781e+154                                                 fv = 21.74893                                                 fv = 7.07213e+14                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 1.340781e+154                                                 fv = 23.08925                                                 fv = 1.340781e+154                                                 fv = 7.720016                                                 fv = 6633655                                                 fv = 1.340781e+154                                                 fv = 6.82204                                                 fv = 1.340781e+154                                                 fv = 25.05257                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.299737                                                 fv = 61.46649                                                 fv = 27.99877                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 6.607756                                                 fv = 3013.961                                                 fv = 710.1547                                                 fv = 53.46735                                                 fv = 5.906197                                                 fv = 1.340781e+154                                                 fv = 59.94934                                                 fv = 9.516841                                                 fv = 39.18184                                                 fv = 25.05257                                                 fv = 1.340781e+154                                                 fv = 1514.958                                                 fv = 6.902551                                                 fv = 5.767372                                                 fv = 6.459118                                                 fv = 7.413592                                                 fv = 7.482753                                                 fv = 4.506868                                                 fv = 5.702409                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 467.4449                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 4.506868                                                 fv = 6.463558                                                 fv = 1751.239 run # 1: f=4.506868216
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.19377                                                 fv = 32.95835                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 312.3735                                                 fv = 17.19377                                                 fv = 1.340781e+154                                                 fv = 26.39736                                                 fv = 19867905032                                                 fv = 967.1852                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.19377                                                 fv = 1.340781e+154                                                 fv = 16.30446                                                 fv = 25.89994                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.30446                                                 fv = 271671.1                                                 fv = 1.340781e+154                                                 fv = 103729.1                                                 fv = 1.340781e+154                                                 fv = 34.56686                                                 fv = 1.340781e+154                                                 fv = 15.29109                                                 fv = 13886.28                                                 fv = 1.340781e+154                                                 fv = 17.7769                                                 fv = 1.340781e+154                                                 fv = 9.441185                                                 fv = 36.70065                                                 fv = 1.340781e+154                                                 fv = 36.70065                                                 fv = 23.98017                                                 fv = 583.5937                                                 fv = 111.2631                                                 fv = 1.340781e+154                                                 fv = 23.98017                                                 fv = 1.340781e+154                                                 fv = 9.778902                                                 fv = 9.441185                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.13288                                                 fv = 204.86                                                 fv = 1.340781e+154                                                 fv = 671.9845                                                 fv = 9.441185                                                 fv = 753.5096                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.83438                                                 fv = 13.83438                                                 fv = 2949.347                                                 fv = 16.30205                                                 fv = 16.30205                                                 fv = 12.98793                                                 fv = 9.441185                                                 fv = 1.340781e+154                                                 fv = 9.441185                                                 fv = 16.22386                                                 fv = 37.87364                                                 fv = 183.8289                                                 fv = 9.441185                                                 fv = 32.77562                                                 fv = 49.30387                                                 fv = 16.30205                                                 fv = 1.340781e+154                                                 fv = 715.2214                                                 fv = 11.81384                                                 fv = 8.672718                                                 fv = 1.340781e+154                                                 fv = 78.25454                                                 fv = 11.81384                                                 fv = 52.24939                                                 fv = 4.371812                                                 fv = 1015.25                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.30107                                                 fv = 4.371812                                                 fv = 22.30509                                                 fv = 10.30107                                                 fv = 1.340781e+154                                                 fv = 12.6572                                                 fv = 4.371812                                                 fv = 1.340781e+154                                                 fv = 74.44013                                                 fv = 4.371812                                                 fv = 1.340781e+154                                                 fv = 24.25493                                                 fv = 18.36678                                                 fv = 17.26359                                                 fv = 17.25768                                                 fv = 1.340781e+154                                                 fv = 9.389457                                                 fv = 1.340781e+154                                                 fv = 277.605                                                 fv = 11.81384                                                 fv = 17.98792                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.292609                                                 fv = 5.102348                                                 fv = 13.83001                                                 fv = 64.95467                                                 fv = 1.340781e+154                                                 fv = 93.14188                                                 fv = 7.777073                                                 fv = 4.371812                                                 fv = 9.994986                                                 fv = 4.371812                                                 fv = 4.371812                                                 fv = 4.371812                                                 fv = 4.371812                                                 fv = 1.340781e+154                                                 fv = 152.4538                                                 fv = 11.09428 run # 2: f=4.371812157
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.91996                                                 fv = 32.95835                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.80193                                                 fv = 23.91996                                                 fv = 28.17384                                                 fv = 936.191                                                 fv = 3181190                                                 fv = 5913.801                                                 fv = 34.85553                                                 fv = 1668.581                                                 fv = 6.970034                                                 fv = 1.340781e+154                                                 fv = 35.92897                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 126.9499                                                 fv = 4602.977                                                 fv = 4602.977                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.970034                                                 fv = 1.340781e+154                                                 fv = 4602.977                                                 fv = 6.970034                                                 fv = 619.5773                                                 fv = 6.970034                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.108919                                                 fv = 119.3979                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.29115                                                 fv = 368.1798                                                 fv = 6.970034                                                 fv = 1.340781e+154                                                 fv = 67.88458                                                 fv = 312741.2                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26255.46                                                 fv = 13713.7                                                 fv = 2001.408                                                 fv = 881.8661                                                 fv = 24115.12                                                 fv = 1001.648                                                 fv = 1.340781e+154                                                 fv = 1530.881                                                 fv = 1.340781e+154                                                 fv = 10.53133                                                 fv = 52837.27                                                 fv = 9.814353                                                 fv = 4658.072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.913935                                                 fv = 17.30344                                                 fv = 4.710591                                                 fv = 993239307                                                 fv = 1.340781e+154                                                 fv = 8.960515                                                 fv = 25.52081                                                 fv = 11.85065                                                 fv = 30.12243                                                 fv = 1.340781e+154                                                 fv = 4.710591                                                 fv = 7.536673                                                 fv = 20.20654                                                 fv = 1.340781e+154                                                 fv = 9.093983                                                 fv = 27.46213                                                 fv = 8.989381                                                 fv = 8.640358                                                 fv = 4.710591                                                 fv = 1890.403                                                 fv = 4.710591                                                 fv = 9.333815                                                 fv = 1.340781e+154                                                 fv = 21.77561                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.710591                                                 fv = 4.710591                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.5643                                                 fv = 29521.37                                                 fv = 29.52777                                                 fv = 4.727006                                                 fv = 4.5643                                                 fv = 6.168821                                                 fv = 4.5643                                                 fv = 1.340781e+154                                                 fv = 4.5643                                                 fv = 4.5643                                                 fv = 4.5643                                                 fv = 29.11756                                                 fv = 1727.268                                                 fv = 31.78709                                                 fv = 1.340781e+154                                                 fv = 4.5643                                                 fv = 21.8697                                                 fv = 1.340781e+154                                                 fv = 1157.511                                                 fv = 5.069452                                                 fv = 5.998048                                                 fv = 6.832753                                                 fv = 37.61133 run # 4: f=4.564300325
#> 
#> bb eval : 601
#> best    : 4.258467182
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 2 2 0 1 0 1 0 0 3 1 1 2 1 1 1 2 1 1 ) f(x) = 4.258467182
#> 
#>                                                 fv = 4.258467                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.954703