Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 15.16977                                                 fv = 59.83249                                                 fv = 18.0705                                                 fv = 25.1252                                                 fv = 10400.64                                                 fv = 8.376867                                                 fv = 67077747222                                                 fv = 1.340781e+154                                                 fv = 10.22528                                                 fv = 33.48994                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.53736                                                 fv = 14.05371                                                 fv = 10.35299                                                 fv = 8.376867                                                 fv = 14.05371                                                 fv = 14.05371                                                 fv = 83748.9                                                 fv = 17210.03                                                 fv = 33.48994                                                 fv = 39.65681                                                 fv = 12.32556                                                 fv = 1.340781e+154                                                 fv = 8.807638                                                 fv = 8.376867                                                 fv = 1219.074                                                 fv = 1.340781e+154                                                 fv = 2546.636                                                 fv = 1.340781e+154                                                 fv = 8.376867                                                 fv = 14.05371                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.39131                                                 fv = 12.52897                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.32013                                                 fv = 44.12363                                                 fv = 5961.756                                                 fv = 678.7074                                                 fv = 47.00175                                                 fv = 138.5609                                                 fv = 1.340781e+154                                                 fv = 7.17565                                                 fv = 1.340781e+154                                                 fv = 31.9803                                                 fv = 10.98819                                                 fv = 7.589094                                                 fv = 6.923904                                                 fv = 1.340781e+154                                                 fv = 55.90638                                                 fv = 18.05188                                                 fv = 24.06876                                                 fv = 23.06809                                                 fv = 18.45959                                                 fv = 29.01683                                                 fv = 8.287145                                                 fv = 1.340781e+154                                                 fv = 5.671334                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.391206                                                 fv = 8.238647                                                 fv = 24.5148                                                 fv = 8.506486                                                 fv = 7.428301                                                 fv = 33.73673                                                 fv = 5.671334                                                 fv = 1.340781e+154                                                 fv = 8.813926                                                 fv = 5.671334                                                 fv = 7.435665                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 231.2783                                                 fv = 8.548308                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 317.8296                                                 fv = 21.38441                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.79815                                                 fv = 1.340781e+154                                                 fv = 43.55026                                                 fv = 5.671334                                                 fv = 19.3469                                                 fv = 8.640215                                                 fv = 8.640215                                                 fv = 46.49133                                                 fv = 51.12754                                                 fv = 7.222663                                                 fv = 5.231786                                                 fv = 1.340781e+154                                                 fv = 6.909781                                                 fv = 14.39828                                                 fv = 17.03808                                                 fv = 28.52223                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.545333                                                 fv = 61.48236                                                 fv = 5.749359                                                 fv = 6.158563                                                 fv = 5.320203                                                 fv = 5.320203                                                 fv = 5.626728                                                 fv = 12.17487                                                 fv = 10.2517                                                 fv = 8.57961                                                 fv = 5.657552                                                 fv = 6.288408                                                 fv = 299.0225                                                 fv = 1.340781e+154                                                 fv = 13.37603                                                 fv = 5.231786                                                 fv = 5.231786                                                 fv = 6.122754                                                 fv = 6.057253                                                 fv = 1.340781e+154                                                 fv = 6.490279                                                 fv = 5.231786                                                 fv = 1.340781e+154                                                 fv = 6.374984                                                 fv = 6.05174                                                 fv = 5.115069                                                 fv = 5.656174                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.115069                                                 fv = 5.115069                                                 fv = 5.115069                                                 fv = 1.340781e+154                                                 fv = 7.255257                                                 fv = 1.340781e+154                                                 fv = 5.739133                                                 fv = 6.601934                                                 fv = 9.927702                                                 fv = 7.271589                                                 fv = 5.656174                                                 fv = 6.736001                                                 fv = 1.340781e+154                                                 fv = 6.57741                                                 fv = 6.251288                                                 fv = 7.33752                                                 fv = 9.57087                                                 fv = 12.28926                                                 fv = 10.31014                                                 fv = 46.04598 run # 0: f=5.115068767
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.87725                                                 fv = 47.00175                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 47.00175                                                 fv = 47.00175                                                 fv = 47.00175                                                 fv = 68.56589                                                 fv = 15.87725                                                 fv = 9.850563                                                 fv = 164.7827                                                 fv = 11.27383                                                 fv = 12.36017                                                 fv = 5.336642                                                 fv = 9.522755                                                 fv = 30.38237                                                 fv = 30.38237                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.336642                                                 fv = 5.336642                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.97145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.689082                                                 fv = 172.3886                                                 fv = 134.4295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.90998                                                 fv = 639.9826                                                 fv = 1.340781e+154                                                 fv = 113.0993                                                 fv = 1.340781e+154                                                 fv = 9.957918                                                 fv = 28.46255                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 30.38237                                                 fv = 92.96726                                                 fv = 1051359                                                 fv = 5.336642                                                 fv = 7.153915                                                 fv = 9.580405                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.45914                                                 fv = 1.340781e+154                                                 fv = 1428905765                                                 fv = 5.336642                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.541074                                                 fv = 1.340781e+154                                                 fv = 9.471325                                                 fv = 8.450564                                                 fv = 7863.547                                                 fv = 6.052277                                                 fv = 18.81248                                                 fv = 221682.3                                                 fv = 19.11766                                                 fv = 6.052277                                                 fv = 6.052277                                                 fv = 5.336642                                                 fv = 6.846412                                                 fv = 7.616844                                                 fv = 1.340781e+154                                                 fv = 5.336642                                                 fv = 1.340781e+154                                                 fv = 378.2808                                                 fv = 8.018555                                                 fv = 1.340781e+154                                                 fv = 20.56303 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 10.02307                                                 fv = 5.336642                                                 fv = 1057.759                                                 fv = 5.336642                                                 fv = 5.336642                                                 fv = 5.336642                                                 fv = 5.00589                                                 fv = 13.3453                                                 fv = 1.340781e+154                                                 fv = 7.161403                                                 fv = 17.27802                                                 fv = 6.364348                                                 fv = 1.340781e+154                                                 fv = 6.178411                                                 fv = 5.292055                                                 fv = 9.117984                                                 fv = 6.178411                                                 fv = 5.00589                                                 fv = 6.428262                                                 fv = 6.014445                                                 fv = 5.00589                                                 fv = 9.533905                                                 fv = 1.340781e+154                                                 fv = 5.336642                                                 fv = 5.436302                                                 fv = 6.598776                                                 fv = 12.43385                                                 fv = 1.340781e+154                                                 fv = 6.308241                                                 fv = 5.00589                                                 fv = 13.13161                                                 fv = 18673626                                                 fv = 1.340781e+154                                                 fv = 5.564896                                                 fv = 5.610856                                                 fv = 1.340781e+154                                                 fv = 5.970577                                                 fv = 6.665966                                                 fv = 5.00589                                                 fv = 4.869944                                                 fv = 7.081559                                                 fv = 1.340781e+154                                                 fv = 5.00589                                                 fv = 1.340781e+154                                                 fv = 6.107787                                                 fv = 6.782525                                                 fv = 7.914282                                                 fv = 6.687636                                                 fv = 6.396118                                                 fv = 6.182344                                                 fv = 5.336642                                                 fv = 4.869944                                                 fv = 4.869944                                                 fv = 6.407575                                                 fv = 5.412783                                                 fv = 6.054987                                                 fv = 9.349722                                                 fv = 6.043916                                                 fv = 5.23925                                                 fv = 7.132785                                                 fv = 1.340781e+154                                                 fv = 8.219711                                                 fv = 4.869944                                                 fv = 6.613443                                                 fv = 4.869944                                                 fv = 4.869944                                                 fv = 5.481451                                                 fv = 8.324714                                                 fv = 4.869944                                                 fv = 4.869944                                                 fv = 4.869944                                                 fv = 12.51143                                                 fv = 18.83781 run # 1: f=4.869944215
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 144386.4                                                 fv = 47.00175                                                 fv = 47.00175                                                 fv = 47.00175                                                 fv = 42.31609                                                 fv = 606.128                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.61154                                                 fv = 18.61154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.61154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.37344                                                 fv = 1379716647                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.90262                                                 fv = 1.340781e+154                                                 fv = 22.5914                                                 fv = 33.58262                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.20697                                                 fv = 11.45201                                                 fv = 1.340781e+154                                                 fv = 23.16973                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 109.229                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 125.3444                                                 fv = 11.31039                                                 fv = 61.63546                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 1.340781e+154                                                 fv = 75.02182                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.60808                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.60808                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 74.31584                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5501.29                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.27493                                                 fv = 53.27493                                                 fv = 57.32743                                                 fv = 1.340781e+154                                                 fv = 223.863                                                 fv = 48.98263                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 59.07752                                                 fv = 11.31039                                                 fv = 916.6818                                                 fv = 17.43511                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 61.63546                                                 fv = 11.31039                                                 fv = 11.31039                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 11.31039                                                 fv = 11.31039                                                 fv = 19.13437                                                 fv = 189.1079                                                 fv = 25.363                                                 fv = 189.1079                                                 fv = 1.340781e+154                                                 fv = 1797.075                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 83.82466                                                 fv = 18.2184                                                 fv = 18.96501                                                 fv = 11.31039                                                 fv = 11.89667                                                 fv = 11.31039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 44.56049                                                 fv = 44.56049                                                 fv = 119.9467                                                 fv = 1.340781e+154                                                 fv = 935534.2                                                 fv = 3439.021                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42.35542                                                 fv = 1.340781e+154                                                 fv = 317.1938                                                 fv = 32.62504                                                 fv = 5280.643                                                 fv = 11.31039                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 46.7947                                                 fv = 1.340781e+154                                                 fv = 11.31039                                                 fv = 11.31039                                                 fv = 11.31039                                                 fv = 9.594112                                                 fv = 2034.067                                                 fv = 1.340781e+154                                                 fv = 2186.222                                                 fv = 27.69296                                                 fv = 14.36799                                                 fv = 9.594112                                                 fv = 9.594112                                                 fv = 4.072516                                                 fv = 25.35761                                                 fv = 1.340781e+154                                                 fv = 2186.222                                                 fv = 19.36846                                                 fv = 60.94893                                                 fv = 4.072516                                                 fv = 19.69087                                                 fv = 4.072516                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24522.98                                                 fv = 4.072516                                                 fv = 21.2062                                                 fv = 1.340781e+154                                                 fv = 4.072516                                                 fv = 48.36054                                                 fv = 5.780663                                                 fv = 4.789069                                                 fv = 4.072516                                                 fv = 17.28321                                                 fv = 1.340781e+154                                                 fv = 1042.716                                                 fv = 1.340781e+154 run # 2: f=4.072516131
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 625
#> best    : 4.072516131
#> worst   : 1.340780793e+154
#> solution: x = ( 0 3 0 0 1 2 3 3 0 0 1 1 9 4 1 1 2 2 7 5 ) f(x) = 4.072516131
#> 
#>                                                 fv = 4.072516                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (3): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 16969.41