Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 28.27872                                                 fv = 31.55425                                                 fv = 8.030612                                                 fv = 3785155539                                                 fv = 1.340781e+154                                                 fv = 88.73417                                                 fv = 683.7873                                                 fv = 297942608046                                                 fv = 1.340781e+154                                                 fv = 72707.6                                                 fv = 14943.83                                                 fv = 1.340781e+154                                                 fv = 54829.81                                                 fv = 1.340781e+154                                                 fv = 2820.502                                                 fv = 1.340781e+154                                                 fv = 683.7873                                                 fv = 683.7873                                                 fv = 1.340781e+154                                                 fv = 55914.17                                                 fv = 1.340781e+154                                                 fv = 75.34587                                                 fv = 155755681                                                 fv = 21.08883                                                 fv = 6691043                                                 fv = 8.540813                                                 fv = 4065.226                                                 fv = 1.340781e+154                                                 fv = 282.539                                                 fv = 12.6266                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.030612                                                 fv = 12.6266                                                 fv = 8.030612                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.40689                                                 fv = 35.79296                                                 fv = 1.340781e+154                                                 fv = 21936.72                                                 fv = 1517.665                                                 fv = 30.77825                                                 fv = 12.95572                                                 fv = 1.340781e+154                                                 fv = 42.01793                                                 fv = 10.04205                                                 fv = 1.340781e+154                                                 fv = 8.030612                                                 fv = 8.030612                                                 fv = 8.030612                                                 fv = 8.030612                                                 fv = 1.340781e+154                                                 fv = 14.17828                                                 fv = 13.12582                                                 fv = 1.340781e+154                                                 fv = 716.8409                                                 fv = 15.92493                                                 fv = 22125.3                                                 fv = 17.22156                                                 fv = 8.030612                                                 fv = 8.481965 run # 0: f=8.030611728
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.52841                                                 fv = 30.77825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.32345                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9894.998                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 185253563                                                 fv = 260973492                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.24919                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 11.90983                                                 fv = 10.52841                                                 fv = 10.52841                                                 fv = 4457339                                                 fv = 10.52841                                                 fv = 44548.61                                                 fv = 1.340781e+154                                                 fv = 42.1825                                                 fv = 1.340781e+154                                                 fv = 20.72011                                                 fv = 25.35025                                                 fv = 1.340781e+154                                                 fv = 12657.82                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 114.3                                                 fv = 1.340781e+154                                                 fv = 30.77825                                                 fv = 14.35749                                                 fv = 14.09271                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.28124                                                 fv = 16.30963                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 10.52841                                                 fv = 10.52841                                                 fv = 37.94869                                                 fv = 44.34577                                                 fv = 1.340781e+154                                                 fv = 15.74033                                                 fv = 37.94869                                                 fv = 18.27396                                                 fv = 59.52591                                                 fv = 10.52841                                                 fv = 1.340781e+154                                                 fv = 22.48671                                                 fv = 10.29272                                                 fv = 1.340781e+154                                                 fv = 30.77825                                                 fv = 13.89144                                                 fv = 10.29272                                                 fv = 10.29272                                                 fv = 6.853778                                                 fv = 96173.82                                                 fv = 1.340781e+154                                                 fv = 30.77825                                                 fv = 894.548                                                 fv = 6.853778                                                 fv = 144.8247                                                 fv = 444.0581                                                 fv = 13.29737                                                 fv = 1.340781e+154                                                 fv = 6.853778                                                 fv = 10.78994                                                 fv = 6.950427                                                 fv = 6.853778                                                 fv = 6.853778                                                 fv = 26.91623                                                 fv = 16.85317                                                 fv = 144.8247                                                 fv = 30.77825                                                 fv = 6.853778                                                 fv = 178.5625                                                 fv = 19.7453                                                 fv = 17.61913                                                 fv = 13.29737                                                 fv = 15791571620                                                 fv = 1.340781e+154                                                 fv = 14.84537                                                 fv = 9.236869                                                 fv = 6.853778                                                 fv = 6.853778                                                 fv = 261.1434                                                 fv = 6.853778                                                 fv = 16.18322                                                 fv = 27.11098                                                 fv = 61.39182                                                 fv = 6.853778                                                 fv = 6.853778                                                 fv = 158.7413                                                 fv = 6.853778                                                 fv = 30.61512                                                 fv = 7.401948                                                 fv = 144.8247                                                 fv = 1.340781e+154                                                 fv = 144.8247                                                 fv = 15.3052                                                 fv = 16.9028                                                 fv = 20.92992                                                 fv = 46103.67                                                 fv = 158.7413                                                 fv = 7.611575                                                 fv = 6.964849                                                 fv = 6.853778                                                 fv = 6.853778                                                 fv = 6.853778                                                 fv = 25.03106 run # 1: f=6.853777613
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.06566                                                 fv = 30.77825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.45704                                                 fv = 15.18062                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.22221                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.97989                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.77884                                                 fv = 17.13579                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.06566                                                 fv = 1.340781e+154                                                 fv = 8304.227                                                 fv = 5.86368                                                 fv = 18.17944                                                 fv = 7.648601                                                 fv = 24.81437 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 15.73611                                                 fv = 3.823899                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.12259                                                 fv = 16.83675                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.823899                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.587849                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.42914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.51735                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.81399                                                 fv = 1.340781e+154                                                 fv = 5840993                                                 fv = 1.340781e+154                                                 fv = 74473.36                                                 fv = 1.340781e+154                                                 fv = 356322.5                                                 fv = 3.823899                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 216.8369                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.823899                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.419461                                                 fv = 6325.964                                                 fv = 21.07426                                                 fv = 7.149926                                                 fv = 1.340781e+154                                                 fv = 34.20023                                                 fv = 4.813677                                                 fv = 22.07853                                                 fv = 5401.961                                                 fv = 1.340781e+154                                                 fv = 15.26796                                                 fv = 105.6677                                                 fv = 3.823899                                                 fv = 5.235873                                                 fv = 1.340781e+154                                                 fv = 12.125                                                 fv = 1.340781e+154                                                 fv = 8.094333                                                 fv = 3.823899                                                 fv = 1.340781e+154                                                 fv = 217.567                                                 fv = 106.6428                                                 fv = 3.823899                                                 fv = 14.35749 run # 2: f=3.823899466
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19837.93                                                 fv = 30.77825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 30.77825                                                 fv = 38.23266                                                 fv = 17.87977                                                 fv = 1877.009                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.007899e+12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.75519                                                 fv = 76.1551                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 56.69769                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3511052974                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.91582                                                 fv = 162.6216                                                 fv = 19464.41                                                 fv = 1.340781e+154                                                 fv = 21.17451                                                 fv = 16.95025                                                 fv = 23.91155                                                 fv = 889324.5                                                 fv = 22.91582                                                 fv = 159901839                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.79722                                                 fv = 22.91582                                                 fv = 15.79722                                                 fv = 1.340781e+154                                                 fv = 40.18971                                                 fv = 455.5657                                                 fv = 1.340781e+154                                                 fv = 177472.2                                                 fv = 20.58095                                                 fv = 168.4251                                                 fv = 531.3886                                                 fv = 15.75519                                                 fv = 124.9762                                                 fv = 96.84126                                                 fv = 95.94089                                                 fv = 1.340781e+154                                                 fv = 25.86964                                                 fv = 15.75519                                                 fv = 15.75519                                                 fv = 1.340781e+154                                                 fv = 45.26657                                                 fv = 24.87902                                                 fv = 1.340781e+154                                                 fv = 14.11319                                                 fv = 14.11319                                                 fv = 19409.23                                                 fv = 1.340781e+154                                                 fv = 15.65531                                                 fv = 14.11319                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.11319                                                 fv = 22.73631                                                 fv = 20.446                                                 fv = 26.86419                                                 fv = 21.11528                                                 fv = 14.11319                                                 fv = 15.87794                                                 fv = 1.340781e+154                                                 fv = 1371.557                                                 fv = 32.98311                                                 fv = 15.75519                                                 fv = 15.65531                                                 fv = 1.340781e+154                                                 fv = 12.64118                                                 fv = 1.340781e+154                                                 fv = 154.4029                                                 fv = 24.43792                                                 fv = 12.72509                                                 fv = 16.25549                                                 fv = 12.73995                                                 fv = 13.64227                                                 fv = 15.11691                                                 fv = 12.64118                                                 fv = 13.99135                                                 fv = 12.64118                                                 fv = 13.31022                                                 fv = 1.340781e+154                                                 fv = 18.76527                                                 fv = 13.28547                                                 fv = 22.70545                                                 fv = 18.36996                                                 fv = 1.340781e+154                                                 fv = 12.64118                                                 fv = 12.64118                                                 fv = 504.249                                                 fv = 998.1411                                                 fv = 80220316                                                 fv = 1.340781e+154                                                 fv = 15.48359                                                 fv = 10.58534                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 341.0732                                                 fv = 14.19048                                                 fv = 10.58534                                                 fv = 9.1506                                                 fv = 265.7849                                                 fv = 1.340781e+154                                                 fv = 15.59428                                                 fv = 34.73933                                                 fv = 13.88235                                                 fv = 19.96417                                                 fv = 110.5264                                                 fv = 83.91991                                                 fv = 8.493687                                                 fv = 110.5264                                                 fv = 1.340781e+154                                                 fv = 13.80796                                                 fv = 110.5264                                                 fv = 1.340781e+154                                                 fv = 102.1535                                                 fv = 16.17194                                                 fv = 9.1506                                                 fv = 110.5264                                                 fv = 26.76874                                                 fv = 84.3802                                                 fv = 11.05202                                                 fv = 363.0862                                                 fv = 88.41909                                                 fv = 110.5264                                                 fv = 66.11724                                                 fv = 13.67286                                                 fv = 8.493687                                                 fv = 68.88501                                                 fv = 9.918864                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.80796                                                 fv = 8.493687                                                 fv = 13.06098                                                 fv = 44.39716                                                 fv = 42.59237                                                 fv = 11.13715                                                 fv = 1.340781e+154                                                 fv = 7.400868                                                 fv = 5.708854                                                 fv = 1.340781e+154                                                 fv = 12.46634                                                 fv = 48275.15                                                 fv = 6.015685                                                 fv = 10.82295                                                 fv = 11.76494                                                 fv = 6.012076                                                 fv = 6.138751                                                 fv = 5.708854                                                 fv = 5.552143                                                 fv = 1.340781e+154                                                 fv = 10.72364                                                 fv = 6.325253                                                 fv = 8.040484                                                 fv = 7.739088                                                 fv = 6.202213                                                 fv = 6.202213                                                 fv = 5.837894                                                 fv = 8.298999                                                 fv = 1.340781e+154                                                 fv = 5.552143                                                 fv = 73.58404                                                 fv = 5.955843                                                 fv = 8.014117                                                 fv = 5.552143                                                 fv = 6.088299                                                 fv = 5.552143                                                 fv = 10.55749                                                 fv = 5.552143                                                 fv = 6.556053                                                 fv = 2590.506                                                 fv = 62.17267                                                 fv = 1.340781e+154                                                 fv = 8.578414                                                 fv = 11.8426                                                 fv = 6.489838                                                 fv = 5.552143                                                 fv = 9.145514                                                 fv = 15.17548                                                 fv = 6.428242                                                 fv = 5.552143                                                 fv = 7.002103                                                 fv = 5.877407                                                 fv = 12.05386                                                 fv = 8.736281 run # 4: f=5.552143444
#> 
#> bb eval : 626
#> best    : 3.823899466
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 1 2 2 1 1 0 0 0 2 1 1 2 2 2 1 1 2 2 ) f(x) = 3.823899466
#> 
#>                                                 fv = 3.823899                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.441224