Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 11.03451                                                 fv = 22.85074                                                 fv = 10.98811                                                 fv = 566433310                                                 fv = 1.340781e+154                                                 fv = 31.1771                                                 fv = 19.84508                                                 fv = 778.1395                                                 fv = 1.340781e+154                                                 fv = 36.77694                                                 fv = 51.56477                                                 fv = 1.340781e+154                                                 fv = 45.446                                                 fv = 1.340781e+154                                                 fv = 92.06633                                                 fv = 1.340781e+154                                                 fv = 19.84508                                                 fv = 19.84508                                                 fv = 1.340781e+154                                                 fv = 8178.411                                                 fv = 1.340781e+154                                                 fv = 29.37166                                                 fv = 4478.424                                                 fv = 295.6794                                                 fv = 61.98138                                                 fv = 30.25607                                                 fv = 30.25607                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.98811                                                 fv = 11.11545                                                 fv = 10.98811                                                 fv = 10.98811                                                 fv = 58.26844                                                 fv = 157.3967                                                 fv = 86.30803                                                 fv = 388.0656                                                 fv = 124.0234                                                 fv = 21.63067                                                 fv = 56.92887                                                 fv = 62.17934                                                 fv = 12.90429                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.807928                                                 fv = 1.340781e+154                                                 fv = 26.68651                                                 fv = 1242.281                                                 fv = 25.10014                                                 fv = 85.03713                                                 fv = 177.6105                                                 fv = 31.88611                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.63599                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33.61138                                                 fv = 18.93343                                                 fv = 129.1029                                                 fv = 107.1284                                                 fv = 12.6334                                                 fv = 41.81308                                                 fv = 10746.19                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.45765                                                 fv = 34.14734                                                 fv = 149.6716                                                 fv = 49.91852                                                 fv = 28.24036                                                 fv = 5.596317                                                 fv = 50130.85                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 56.16182                                                 fv = 176.381                                                 fv = 65.56925                                                 fv = 7.495504                                                 fv = 3.505843                                                 fv = 3.505843                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6653678                                                 fv = 2606.216                                                 fv = 1.340781e+154                                                 fv = 20.6435                                                 fv = 56443.11                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.505843                                                 fv = 7.397068                                                 fv = 8.60036                                                 fv = 3.505843                                                 fv = 24.51351                                                 fv = 279.6342                                                 fv = 103.5837                                                 fv = 5.791909                                                 fv = 17.94144                                                 fv = 10.90023                                                 fv = 9.335215                                                 fv = 135.8144                                                 fv = 3569.904                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 85.20114                                                 fv = 1.340781e+154                                                 fv = 13.96754                                                 fv = 11.82317                                                 fv = 14.45029                                                 fv = 634.9996                                                 fv = 11.82317                                                 fv = 11.548                                                 fv = 49.4258                                                 fv = 63458.16                                                 fv = 6.108596                                                 fv = 11.82317                                                 fv = 192.2341                                                 fv = 3.505843                                                 fv = 35.28039                                                 fv = 1.340781e+154 run # 0: f=3.505842604
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.598727                                                 fv = 43.1213                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.104572                                                 fv = 9.271102                                                 fv = 11.91338                                                 fv = 28.32576                                                 fv = 15.69293                                                 fv = 1.340781e+154                                                 fv = 5.598727                                                 fv = 10.24543                                                 fv = 9.686576                                                 fv = 1.340781e+154                                                 fv = 65.28754                                                 fv = 7.247018                                                 fv = 1491.463                                                 fv = 7.703691                                                 fv = 342.7052                                                 fv = 18.60973                                                 fv = 6.104572                                                 fv = 1.340781e+154                                                 fv = 6.620918                                                 fv = 1.340781e+154                                                 fv = 13.94876                                                 fv = 9.370132                                                 fv = 5.598727                                                 fv = 7.050235                                                 fv = 1.340781e+154                                                 fv = 9.373401                                                 fv = 1.340781e+154                                                 fv = 10.71081                                                 fv = 6.478887                                                 fv = 5.598727                                                 fv = 1.340781e+154                                                 fv = 820.6605                                                 fv = 5.598727                                                 fv = 8.023976                                                 fv = 5.598727                                                 fv = 134.9851                                                 fv = 8.18225                                                 fv = 6.620918                                                 fv = 20.30124                                                 fv = 28.82047                                                 fv = 43.78462                                                 fv = 8.522169                                                 fv = 7.215717                                                 fv = 43.1213                                                 fv = 8.357776                                                 fv = 5.301166                                                 fv = 8.057093                                                 fv = 43.1213                                                 fv = 6.539322                                                 fv = 17.33587                                                 fv = 5.301166                                                 fv = 5.301166                                                 fv = 14.3426                                                 fv = 24.74751                                                 fv = 10.26861                                                 fv = 5.880325                                                 fv = 5.301166                                                 fv = 1.340781e+154                                                 fv = 7.354642                                                 fv = 1.340781e+154                                                 fv = 5.301166                                                 fv = 5.301166                                                 fv = 5.301166                                                 fv = 7.029929                                                 fv = 26.37651                                                 fv = 6.811678                                                 fv = 5.90051 run # 1: f=5.301165682
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.885512                                                 fv = 43.1213                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.885512                                                 fv = 8.885512                                                 fv = 9.652214                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 76.46934                                                 fv = 8.173921                                                 fv = 290408.2                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17100428                                                 fv = 64.80715                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.508524                                                 fv = 1.340781e+154                                                 fv = 56.38079                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.94163                                                 fv = 1.340781e+154                                                 fv = 14.04606                                                 fv = 7.414436                                                 fv = 9.49465                                                 fv = 28.94137                                                 fv = 17.65168                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.1878                                                 fv = 8.751253                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.469121                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.71203                                                 fv = 1.171131e+13                                                 fv = 7.414436                                                 fv = 7.414436                                                 fv = 7.414436                                                 fv = 1.340781e+154                                                 fv = 47.00044                                                 fv = 56.36399                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.71203                                                 fv = 27.1684                                                 fv = 1.340781e+154                                                 fv = 19.17586                                                 fv = 14.97468                                                 fv = 11.29337                                                 fv = 7.414436                                                 fv = 7.414436                                                 fv = 12.91078                                                 fv = 10.19625                                                 fv = 1.340781e+154                                                 fv = 7.982966                                                 fv = 7.414436                                                 fv = 8.644973                                                 fv = 1.340781e+154                                                 fv = 18.48514                                                 fv = 677.0802                                                 fv = 7.414436                                                 fv = 18.63399                                                 fv = 6.22352                                                 fv = 1.340781e+154                                                 fv = 17.92456                                                 fv = 1.340781e+154                                                 fv = 8.310341 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 6.447223                                                 fv = 5.745391                                                 fv = 35.29521                                                 fv = 1.340781e+154                                                 fv = 8.65049                                                 fv = 5.745391                                                 fv = 14.88565                                                 fv = 5.745391                                                 fv = 10.92266                                                 fv = 7.53076                                                 fv = 10.8846                                                 fv = 10.8846                                                 fv = 10.8846                                                 fv = 11.67773                                                 fv = 28.32576                                                 fv = 28.32576                                                 fv = 26.49309                                                 fv = 7.595106                                                 fv = 32.27345                                                 fv = 11.3694                                                 fv = 28.32576                                                 fv = 9.521329                                                 fv = 803.2836                                                 fv = 1.340781e+154                                                 fv = 13.06416                                                 fv = 7.015163                                                 fv = 5.745391                                                 fv = 5.745391                                                 fv = 1.340781e+154                                                 fv = 10.8846                                                 fv = 10.8846                                                 fv = 5.745391                                                 fv = 18.94742                                                 fv = 12.24949                                                 fv = 11.15804                                                 fv = 10.49924                                                 fv = 10.8846                                                 fv = 10.8846                                                 fv = 11.07175                                                 fv = 28.32576                                                 fv = 10.8846                                                 fv = 6.170002                                                 fv = 58.02569                                                 fv = 30.9067                                                 fv = 6.736483                                                 fv = 8.266223                                                 fv = 9.861426                                                 fv = 12.74535                                                 fv = 1.340781e+154                                                 fv = 16.56341                                                 fv = 5.745391                                                 fv = 11.75195                                                 fv = 5.745391                                                 fv = 8.944014                                                 fv = 5.745391                                                 fv = 5.745391                                                 fv = 5.745391                                                 fv = 5.745391                                                 fv = 6.718775                                                 fv = 6.918122                                                 fv = 9.940009                                                 fv = 9.866751                                                 fv = 10.78587 run # 2: f=5.745391417
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.484363                                                 fv = 43.1213                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.14697                                                 fv = 8.911257                                                 fv = 13.14482                                                 fv = 7.484363                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.484363                                                 fv = 103.4144                                                 fv = 1.340781e+154                                                 fv = 7.484363                                                 fv = 1.340781e+154                                                 fv = 26.49715                                                 fv = 1.340781e+154                                                 fv = 94.51006                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 556.9832                                                 fv = 1.340781e+154                                                 fv = 1199.804                                                 fv = 7.484363                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.921879                                                 fv = 9.344046                                                 fv = 1.340781e+154                                                 fv = 1.370439e+14                                                 fv = 7.484363                                                 fv = 7.484363                                                 fv = 6.83278                                                 fv = 12.22864                                                 fv = 1.340781e+154                                                 fv = 11.37559                                                 fv = 6.83278                                                 fv = 1.340781e+154                                                 fv = 6.936221                                                 fv = 11.79156                                                 fv = 262.8748                                                 fv = 10333.89                                                 fv = 6.936221                                                 fv = 2.506057e+14                                                 fv = 9.898628                                                 fv = 6.83278                                                 fv = 6.83278                                                 fv = 6089.231                                                 fv = 108.7629                                                 fv = 12.14697                                                 fv = 14.73841                                                 fv = 1.340781e+154                                                 fv = 14.0263                                                 fv = 6.83278                                                 fv = 12.14697                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.4653                                                 fv = 1.340781e+154                                                 fv = 6.83278                                                 fv = 7.484363                                                 fv = 9.698164                                                 fv = 7.484363                                                 fv = 7.929831                                                 fv = 9.297328                                                 fv = 9.297328                                                 fv = 6.83278                                                 fv = 1.340781e+154                                                 fv = 137.101                                                 fv = 1.340781e+154                                                 fv = 103.6678                                                 fv = 6.83278                                                 fv = 1.340781e+154                                                 fv = 24.34879                                                 fv = 9.489863                                                 fv = 36305.37                                                 fv = 9.297328                                                 fv = 32.25388                                                 fv = 11.13565                                                 fv = 18.24606                                                 fv = 15.2856                                                 fv = 525.9313                                                 fv = 7.298059                                                 fv = 6.580254                                                 fv = 1.340781e+154                                                 fv = 6.83278                                                 fv = 1.340781e+154                                                 fv = 7.460293                                                 fv = 6.580254                                                 fv = 7.015589                                                 fv = 5.468994                                                 fv = 1.340781e+154                                                 fv = 6.83278                                                 fv = 66.8762                                                 fv = 7.226945                                                 fv = 6.873086                                                 fv = 5.468994                                                 fv = 5.468994                                                 fv = 8.641267                                                 fv = 5.468994                                                 fv = 6.687908                                                 fv = 1.340781e+154                                                 fv = 6.470821                                                 fv = 1.340781e+154                                                 fv = 5.468994                                                 fv = 1.340781e+154                                                 fv = 5.468994                                                 fv = 7.499414                                                 fv = 6.1828                                                 fv = 6.799338                                                 fv = 6.892386                                                 fv = 6.874761                                                 fv = 9.898628                                                 fv = 6.357588 run # 3: f=5.468994164
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.32576                                                 fv = 43.1213                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.32576                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.68158                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.197717                                                 fv = 125.8541                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1925389                                                 fv = 23.76724                                                 fv = 7.197717                                                 fv = 1.340781e+154                                                 fv = 11.03513                                                 fv = 33.9487                                                 fv = 1.340781e+154                                                 fv = 23.76724                                                 fv = 9.416442                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.670545                                                 fv = 8.494178                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 2357941                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.767259                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1037.797                                                 fv = 1.340781e+154                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 529.5729                                                 fv = 2560.068                                                 fv = 7.561684                                                 fv = 1.340781e+154                                                 fv = 26.48432                                                 fv = 6.670545                                                 fv = 149.1287                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 19.31096                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.622819                                                 fv = 1.340781e+154                                                 fv = 19.31096                                                 fv = 1.340781e+154                                                 fv = 204149.8                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 48.05387                                                 fv = 1.340781e+154                                                 fv = 9.259953                                                 fv = 11.21542                                                 fv = 42.19577                                                 fv = 22.94624                                                 fv = 77.84062                                                 fv = 16.71953                                                 fv = 1.340781e+154                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 19.35789                                                 fv = 12.19894                                                 fv = 6.670545                                                 fv = 6.670545                                                 fv = 6.670545                                                 fv = 1.340781e+154                                                 fv = 9.665464                                                 fv = 6.22352                                                 fv = 43.1213                                                 fv = 1.340781e+154                                                 fv = 279.4792                                                 fv = 6.22352                                                 fv = 7.06439                                                 fv = 5.511383                                                 fv = 3689.226                                                 fv = 1.340781e+154                                                 fv = 48.05387                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 8.345725                                                 fv = 6.054475                                                 fv = 14.64134                                                 fv = 11.91338                                                 fv = 6.655665                                                 fv = 7.088788                                                 fv = 5.511383                                                 fv = 6.516157                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 9.521329                                                 fv = 1.340781e+154                                                 fv = 11.73307                                                 fv = 5.511383                                                 fv = 5.964504                                                 fv = 9.244009                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.11169                                                 fv = 24.72022                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.054345                                                 fv = 6.99938                                                 fv = 6.643097                                                 fv = 5.951983                                                 fv = 6.585901                                                 fv = 12.41631                                                 fv = 6.343283                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 7.05171                                                 fv = 5.511383                                                 fv = 6.040862                                                 fv = 16.07042                                                 fv = 8.720397                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 9.101776                                                 fv = 106.017                                                 fv = 8.902763                                                 fv = 7.314249                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.979239                                                 fv = 5.733757                                                 fv = 5.951983                                                 fv = 6.040862                                                 fv = 6.343283                                                 fv = 5.511383                                                 fv = 7.951236                                                 fv = 6.656406                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 7.144175                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 5.511383                                                 fv = 6.296527 run # 4: f=5.511382617
#> 
#> bb eval : 717
#> best    : 3.505842604
#> worst   : 5.745391417
#> solution: x = ( 1 1 2 2 0 1 0 1 1 2 1 2 1 2 2 1 1 3 1 3 ) f(x) = 3.505842604
#> 
#>                                                 fv = 3.505843                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (3): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 158.5099