Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 37.6073                                                 fv = 191.2427                                                 fv = 19.61012                                                 fv = 641381075853                                                 fv = 1.340781e+154                                                 fv = 1003.367                                                 fv = 1848.934                                                 fv = 4.05333e+11                                                 fv = 1.340781e+154                                                 fv = 152402.4                                                 fv = 138709.7                                                 fv = 1.340781e+154                                                 fv = 16049.14                                                 fv = 1.340781e+154                                                 fv = 4253.294                                                 fv = 1.340781e+154                                                 fv = 1848.934                                                 fv = 1848.934                                                 fv = 1.340781e+154                                                 fv = 50866.29                                                 fv = 1.340781e+154                                                 fv = 46.32964                                                 fv = 91412333                                                 fv = 1003.367                                                 fv = 176427850                                                 fv = 21.44983                                                 fv = 41951.71                                                 fv = 34.13226                                                 fv = 1.340781e+154                                                 fv = 16.10649                                                 fv = 217476.7                                                 fv = 1.340781e+154                                                 fv = 63.77592                                                 fv = 2780.064                                                 fv = 7.624023                                                 fv = 7.624023                                                 fv = 1.340781e+154                                                 fv = 8.88743                                                 fv = 40.27903                                                 fv = 23.94009                                                 fv = 15.42002                                                 fv = 1.340781e+154                                                 fv = 8.849907                                                 fv = 7.624023                                                 fv = 1.340781e+154                                                 fv = 7.624023                                                 fv = 13.13864                                                 fv = 44.99791                                                 fv = 7.624023                                                 fv = 15962.03                                                 fv = 628802.3                                                 fv = 81516.6                                                 fv = 1.340781e+154                                                 fv = 7.798274                                                 fv = 8.198545                                                 fv = 8.88743                                                 fv = 13.34355                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.332944                                                 fv = 7.624023                                                 fv = 15.22714                                                 fv = 32.79175                                                 fv = 46.02142                                                 fv = 1.340781e+154                                                 fv = 7.624023                                                 fv = 7.624023                                                 fv = 1.340781e+154                                                 fv = 7.624023                                                 fv = 15.70107                                                 fv = 14.29063                                                 fv = 16.88008                                                 fv = 9.197776                                                 fv = 1.340781e+154                                                 fv = 16.10649                                                 fv = 19.91947                                                 fv = 9.197776                                                 fv = 37.67705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.624023                                                 fv = 11.95798                                                 fv = 13.04624                                                 fv = 11.99705                                                 fv = 8.88743                                                 fv = 7.624023                                                 fv = 7.624023                                                 fv = 7.624023                                                 fv = 1.340781e+154                                                 fv = 8.332944                                                 fv = 7.624023                                                 fv = 7.624023                                                 fv = 10.40576                                                 fv = 12.00774                                                 fv = 5.609404                                                 fv = 5.609404                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.257461                                                 fv = 12.71786                                                 fv = 5.609404                                                 fv = 6.550276                                                 fv = 11.02231                                                 fv = 5.609404                                                 fv = 5.609404                                                 fv = 5.609404                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.609404                                                 fv = 6.075777                                                 fv = 9.176594                                                 fv = 11.99705                                                 fv = 15.42002                                                 fv = 5.801284                                                 fv = 7.399746                                                 fv = 7.624023                                                 fv = 8.88743                                                 fv = 10.7202                                                 fv = 20.63131                                                 fv = 6.782062                                                 fv = 7.318562 run # 0: f=5.609403819
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.482849                                                 fv = 37.67705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.75741                                                 fv = 7.482849                                                 fv = 29.17294                                                 fv = 1.340781e+154                                                 fv = 27.94546                                                 fv = 60268.07                                                 fv = 169.0498                                                 fv = 29.17294                                                 fv = 1.340781e+154                                                 fv = 29.17294                                                 fv = 1.340781e+154                                                 fv = 45291050036                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.064326                                                 fv = 1.340781e+154                                                 fv = 8.064326                                                 fv = 3064544                                                 fv = 1.340781e+154                                                 fv = 9.075347                                                 fv = 7.110515                                                 fv = 29.8781                                                 fv = 6.674025                                                 fv = 20.46942                                                 fv = 127.0749                                                 fv = 1.340781e+154                                                 fv = 19.55315                                                 fv = 6.674025                                                 fv = 6.674025                                                 fv = 8.546039                                                 fv = 15.11727                                                 fv = 6.674025                                                 fv = 11.43256                                                 fv = 6.971841                                                 fv = 11.66109                                                 fv = 1.340781e+154                                                 fv = 8.546039                                                 fv = 60.16744                                                 fv = 1.340781e+154                                                 fv = 9.03675                                                 fv = 824.9518                                                 fv = 1.340781e+154                                                 fv = 1060307                                                 fv = 1.340781e+154                                                 fv = 208778399                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.05942                                                 fv = 6.674025                                                 fv = 1.340781e+154                                                 fv = 274.1687                                                 fv = 7.698785                                                 fv = 29.61736                                                 fv = 451571215                                                 fv = 21.83649                                                 fv = 106.2319                                                 fv = 8.23035                                                 fv = 6.971841                                                 fv = 19.55315                                                 fv = 110.8981                                                 fv = 1.340781e+154                                                 fv = 7.027871                                                 fv = 1.340781e+154                                                 fv = 24.48699                                                 fv = 6.674025                                                 fv = 1.340781e+154                                                 fv = 116945.8                                                 fv = 15.31983                                                 fv = 37.67705                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 16.76754                                                 fv = 17.01138                                                 fv = 495.3726                                                 fv = 6.674025                                                 fv = 7.88271                                                 fv = 6.674025                                                 fv = 1.340781e+154                                                 fv = 6.674025                                                 fv = 8.702072                                                 fv = 5.975142                                                 fv = 7.330021                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 17.08117                                                 fv = 5.975142                                                 fv = 15.57391                                                 fv = 9.603314                                                 fv = 7.094647                                                 fv = 6.055952                                                 fv = 7.094647                                                 fv = 6.966054                                                 fv = 13.52342                                                 fv = 6.950655                                                 fv = 5.975142                                                 fv = 5.683451                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 16.76754                                                 fv = 17.48113                                                 fv = 9.530944                                                 fv = 7.037884                                                 fv = 11.71542                                                 fv = 5.683451                                                 fv = 9.530944                                                 fv = 7.37008                                                 fv = 8.082935                                                 fv = 1.340781e+154                                                 fv = 164.7182                                                 fv = 8.244521                                                 fv = 6.365771                                                 fv = 7.192327                                                 fv = 11.71542                                                 fv = 11.71542                                                 fv = 17.66958                                                 fv = 12.87023                                                 fv = 1.340781e+154                                                 fv = 6.677675                                                 fv = 12.35202                                                 fv = 7.37008                                                 fv = 1.340781e+154                                                 fv = 41.655                                                 fv = 9.341783                                                 fv = 15.12085                                                 fv = 6.170183                                                 fv = 6.425876                                                 fv = 5.683451                                                 fv = 7.531358                                                 fv = 5.683451                                                 fv = 5.683451                                                 fv = 5.683451                                                 fv = 6.228698                                                 fv = 5.683451                                                 fv = 6.228698                                                 fv = 6.508225                                                 fv = 6.411431                                                 fv = 5.683451                                                 fv = 12.05887                                                 fv = 5.683451                                                 fv = 21.48498                                                 fv = 6.15323                                                 fv = 6.703699                                                 fv = 6.183532                                                 fv = 7.925001                                                 fv = 36.07112 run # 1: f=5.683451087
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 423.5312                                                 fv = 37.67705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.17294                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.94546                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.29009                                                 fv = 19.9723                                                 fv = 7.913632                                                 fv = 251394.9                                                 fv = 10.10407                                                 fv = 29464492858                                                 fv = 1.340781e+154                                                 fv = 9.321406                                                 fv = 9.321406                                                 fv = 55333690827                                                 fv = 5.336636                                                 fv = 7.967742                                                 fv = 39.41599                                                 fv = 16.68379                                                 fv = 6.023646e+12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.441466                                                 fv = 1.340781e+154                                                 fv = 5.336636                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.9723                                                 fv = 1.340781e+154                                                 fv = 3973988726                                                 fv = 5.884512                                                 fv = 44.29902                                                 fv = 10.4795                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44415.99                                                 fv = 9.321406                                                 fv = 1.866519e+12                                                 fv = 7.089348                                                 fv = 1.340781e+154                                                 fv = 5.336636                                                 fv = 10404383962                                                 fv = 4183480                                                 fv = 5.336636                                                 fv = 122.8814                                                 fv = 1.340781e+154                                                 fv = 479670.3                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.03818                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.65717                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13247.11                                                 fv = 34.31131                                                 fv = 29.28382                                                 fv = 33.02494                                                 fv = 4.871297                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.871297                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.224273                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.615733                                                 fv = 1.340781e+154                                                 fv = 4.871297                                                 fv = 1.340781e+154                                                 fv = 14.4068                                                 fv = 539287.1                                                 fv = 223.6955                                                 fv = 1.340781e+154                                                 fv = 2205705                                                 fv = 4.871297                                                 fv = 284.8225                                                 fv = 1.340781e+154                                                 fv = 202.848                                                 fv = 13.92082                                                 fv = 6.224273                                                 fv = 1.340781e+154                                                 fv = 10.16767                                                 fv = 21.55336                                                 fv = 5.844125                                                 fv = 1128.492                                                 fv = 5.873601                                                 fv = 4.871297                                                 fv = 4.871297                                                 fv = 7.916549                                                 fv = 5.642406                                                 fv = 1.340781e+154                                                 fv = 5.334254                                                 fv = 4.726674                                                 fv = 1.340781e+154                                                 fv = 901.0662                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 177288                                                 fv = 11.04664                                                 fv = 556.9505                                                 fv = 3.36112                                                 fv = 1.340781e+154                                                 fv = 3.36112                                                 fv = 3.36112                                                 fv = 1.340781e+154                                                 fv = 3.981319                                                 fv = 9.520012                                                 fv = 2.0981                                                 fv = 22.78996                                                 fv = 1.340781e+154                                                 fv = 29.17294                                                 fv = 7.086383                                                 fv = 11.9541                                                 fv = 1.340781e+154                                                 fv = 2.0981                                                 fv = 2.0981                                                 fv = 2.0981                                                 fv = 6.555174                                                 fv = 6.332954                                                 fv = 19.35776                                                 fv = 119.6757                                                 fv = 151.803                                                 fv = 4.545683                                                 fv = 10.06467                                                 fv = 99.72343                                                 fv = 1932259                                                 fv = 2.340002                                                 fv = 6.161213                                                 fv = 4.935208                                                 fv = 2.0981                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.98379                                                 fv = 7792.016                                                 fv = 5.731531                                                 fv = 638.2806                                                 fv = 37.29008                                                 fv = 2.754498                                                 fv = 2.010416                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.070117                                                 fv = 2.305257                                                 fv = 2.010416                                                 fv = 2.995995                                                 fv = 3.789523                                                 fv = 2.601315                                                 fv = 1.340781e+154                                                 fv = 2.010416                                                 fv = 1.966219                                                 fv = 7.668157                                                 fv = 1.340781e+154                                                 fv = 12.72293                                                 fv = 37.72695                                                 fv = 5.74652                                                 fv = 1.966219                                                 fv = 3.966977                                                 fv = 37.72695                                                 fv = 3.582068                                                 fv = 2.435182                                                 fv = 16.45014                                                 fv = 2.448734                                                 fv = 5.74652                                                 fv = 1.966219                                                 fv = 1.966219                                                 fv = 1.340781e+154                                                 fv = 2.010416                                                 fv = 2.414092                                                 fv = 2.287748                                                 fv = 2.831101                                                 fv = 3.598481                                                 fv = 7.809467                                                 fv = 1.966219                                                 fv = 2.862194                                                 fv = 27.84508                                                 fv = 12.91258                                                 fv = 13.0094                                                 fv = 7.602596                                                 fv = 1.966219                                                 fv = 1445.19 run # 2: f=1.966219268
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 56.48663                                                 fv = 37.67705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 708165121083                                                 fv = 12.73969                                                 fv = 11.70777                                                 fv = 54791.51                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.29591                                                 fv = 11.70777                                                 fv = 11.70777                                                 fv = 10.17695                                                 fv = 1405.96                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.71496                                                 fv = 10.87662                                                 fv = 7.677071                                                 fv = 109987.2                                                 fv = 271.2553                                                 fv = 1.340781e+154                                                 fv = 344.9772                                                 fv = 12.21407                                                 fv = 12.52896                                                 fv = 10.25434                                                 fv = 7.677071                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37986.53                                                 fv = 12.16731                                                 fv = 6.869748e+12                                                 fv = 13.59309                                                 fv = 48.06357                                                 fv = 1.340781e+154                                                 fv = 150138.5                                                 fv = 12.21407                                                 fv = 17.38981                                                 fv = 11.0641                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.991318                                                 fv = 12.0701                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.31821                                                 fv = 16.27009                                                 fv = 14.88142                                                 fv = 13.91508                                                 fv = 1.340781e+154                                                 fv = 11.18189                                                 fv = 11.18189                                                 fv = 11.18189                                                 fv = 1.340781e+154                                                 fv = 12.19444                                                 fv = 11.57113                                                 fv = 44565.63                                                 fv = 116.188                                                 fv = 1.340781e+154                                                 fv = 180849528637                                                 fv = 116.188                                                 fv = 1.340781e+154                                                 fv = 16.21239                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.38109                                                 fv = 17.0504                                                 fv = 10.97439                                                 fv = 17.0504                                                 fv = 1197.807                                                 fv = 1.340781e+154                                                 fv = 15.9722                                                 fv = 12.66789                                                 fv = 18.60637                                                 fv = 9.983132                                                 fv = 1.396079e+14                                                 fv = 1.340781e+154                                                 fv = 9.317374                                                 fv = 10.97439                                                 fv = 1.340781e+154                                                 fv = 15.13041                                                 fv = 15.13041                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.57892                                                 fv = 11.57113                                                 fv = 131.5116                                                 fv = 17.0504                                                 fv = 19.56812                                                 fv = 16.04218                                                 fv = 6.991318                                                 fv = 31.52654                                                 fv = 6.991318                                                 fv = 1.340781e+154                                                 fv = 9.28539                                                 fv = 9.28539                                                 fv = 24800.12                                                 fv = 12.52073                                                 fv = 53.90353                                                 fv = 69617.74                                                 fv = 508349.9                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.53547                                                 fv = 15.23741                                                 fv = 14.11177                                                 fv = 9.703454                                                 fv = 6.34793                                                 fv = 5.95259                                                 fv = 1.340781e+154                                                 fv = 16.53547                                                 fv = 15.23741                                                 fv = 6.422092                                                 fv = 5.95259                                                 fv = 14.05197                                                 fv = 5.95259                                                 fv = 11.12971                                                 fv = 14.83212                                                 fv = 14.05197                                                 fv = 15.65229                                                 fv = 1.340781e+154                                                 fv = 7.305247                                                 fv = 14.73488                                                 fv = 12.45777                                                 fv = 5.95259                                                 fv = 12.45777                                                 fv = 12.63595                                                 fv = 11.016                                                 fv = 7.980169                                                 fv = 154.1375                                                 fv = 7.247127                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.04829                                                 fv = 8.912686                                                 fv = 1335.074                                                 fv = 9.103915                                                 fv = 17.0504                                                 fv = 6.927306                                                 fv = 14.95458                                                 fv = 15.23741                                                 fv = 8.965822                                                 fv = 16.61671                                                 fv = 5.95259                                                 fv = 12.01338                                                 fv = 13.32574                                                 fv = 5.95259                                                 fv = 7.15818                                                 fv = 7.128508                                                 fv = 5.95259                                                 fv = 5.95259                                                 fv = 5.95259                                                 fv = 6.907268                                                 fv = 6.672467                                                 fv = 6.995426                                                 fv = 6.156666                                                 fv = 1.340781e+154                                                 fv = 55.31673 run # 3: f=5.95258977
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 742
#> best    : 1.966219268
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 0 3 0 1 0 1 0 0 1 1 10 1 1 1 1 5 1 3 ) f(x) = 1.966219268
#> 
#>                                                 fv = 1.966219                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.721674