Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 21.20982                                                 fv = 26.27601                                                 fv = 15.08084                                                 fv = 1084.291                                                 fv = 1.340781e+154                                                 fv = 7.895124                                                 fv = 15.19002                                                 fv = 1.340781e+154                                                 fv = 10.18891                                                 fv = 1.340781e+154                                                 fv = 9.22174                                                 fv = 1.340781e+154                                                 fv = 7.416052                                                 fv = 1230.048                                                 fv = 15.87344                                                 fv = 15.87344                                                 fv = 1.340781e+154                                                 fv = 27.98352                                                 fv = 1.340781e+154                                                 fv = 13.06113                                                 fv = 1.340781e+154                                                 fv = 33.60223                                                 fv = 1.340781e+154                                                 fv = 239.2555                                                 fv = 9.82688                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.34013                                                 fv = 12.644                                                 fv = 14.57871                                                 fv = 7.354064e+14                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7239.128                                                 fv = 1.340781e+154                                                 fv = 20.67692                                                 fv = 10.51941                                                 fv = 7.416052                                                 fv = 7.416052                                                 fv = 31.83048                                                 fv = 57.14098                                                 fv = 7.416052                                                 fv = 186.4475                                                 fv = 10.41138                                                 fv = 129.4906                                                 fv = 1.340781e+154                                                 fv = 16.49957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.05452                                                 fv = 130.1763                                                 fv = 1028016                                                 fv = 7.416052                                                 fv = 10.71192                                                 fv = 9.931208                                                 fv = 1.340781e+154                                                 fv = 13.69607                                                 fv = 1.340781e+154                                                 fv = 21.89296                                                 fv = 1.340781e+154                                                 fv = 8.112731                                                 fv = 7.398708                                                 fv = 664.86                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.237712                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.12532                                                 fv = 1.340781e+154                                                 fv = 8.301257                                                 fv = 1.340781e+154                                                 fv = 13.06113                                                 fv = 16.08958                                                 fv = 51.82305                                                 fv = 9.553006                                                 fv = 6.68547                                                 fv = 21.65231                                                 fv = 1862.253                                                 fv = 12.05648                                                 fv = 17.28556                                                 fv = 9.462053                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.237712                                                 fv = 14.77833                                                 fv = 6.405042                                                 fv = 6.237712                                                 fv = 6.971256                                                 fv = 6.237712                                                 fv = 59.45646                                                 fv = 101.8915                                                 fv = 1.340781e+154                                                 fv = 482.8861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1404.851                                                 fv = 6.415219                                                 fv = 9.594197                                                 fv = 1.340781e+154                                                 fv = 8.35104                                                 fv = 35.38536                                                 fv = 75.34915                                                 fv = 26.73081                                                 fv = 1.340781e+154                                                 fv = 15.99651                                                 fv = 19.12831                                                 fv = 6.237712                                                 fv = 6.237712                                                 fv = 6.237712                                                 fv = 16.5213                                                 fv = 8.651297 run # 0: f=6.237712139
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.712299                                                 fv = 30.7943                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.793122                                                 fv = 8.793122                                                 fv = 6.123286                                                 fv = 10.21513                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45936.71                                                 fv = 1869631                                                 fv = 7.116209e+12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2123.767                                                 fv = 6.123286                                                 fv = 10.98671                                                 fv = 10.98671                                                 fv = 18.73173                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.991154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 85.56936                                                 fv = 5.991154                                                 fv = 12.08138                                                 fv = 1.340781e+154                                                 fv = 798817.7                                                 fv = 1.340781e+154                                                 fv = 29370160783                                                 fv = 7.448779                                                 fv = 5.991154                                                 fv = 11.92658                                                 fv = 8.705392                                                 fv = 23958.44                                                 fv = 1240947594                                                 fv = 8.705392                                                 fv = 8.271333                                                 fv = 8.847605                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1903408                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.423845                                                 fv = 1.340781e+154                                                 fv = 9.623321                                                 fv = 1.340781e+154                                                 fv = 10.59115                                                 fv = 15.04644                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.260365                                                 fv = 1.340781e+154                                                 fv = 7.872376                                                 fv = 21.37502                                                 fv = 26341.16                                                 fv = 1.340781e+154                                                 fv = 9.019569                                                 fv = 5.991154                                                 fv = 1.340781e+154                                                 fv = 16.41079                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.991154                                                 fv = 6.031947                                                 fv = 1.340781e+154                                                 fv = 8.828765                                                 fv = 7.810328                                                 fv = 5.979165                                                 fv = 1.340781e+154                                                 fv = 11.91432                                                 fv = 12.84342                                                 fv = 5.700075                                                 fv = 291.1491                                                 fv = 8.286615                                                 fv = 8.939738                                                 fv = 1.340781e+154                                                 fv = 15.82956                                                 fv = 5.700075                                                 fv = 6.374795                                                 fv = 5.714673                                                 fv = 5.700075                                                 fv = 5.700075                                                 fv = 5.700075                                                 fv = 5.678631                                                 fv = 4.908543                                                 fv = 134.8627 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 12.45592                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.908543                                                 fv = 1.340781e+154                                                 fv = 7.812257                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.908543                                                 fv = 4.908543                                                 fv = 6.531685                                                 fv = 5.872049                                                 fv = 1.340781e+154                                                 fv = 8.473393                                                 fv = 1.340781e+154                                                 fv = 6.939585                                                 fv = 10.52505                                                 fv = 81.71663                                                 fv = 51.81204                                                 fv = 6.946649                                                 fv = 133.1566                                                 fv = 1.340781e+154                                                 fv = 302.0764                                                 fv = 12.59678                                                 fv = 1.340781e+154                                                 fv = 7.256849                                                 fv = 10.38293                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.007865                                                 fv = 9.489105                                                 fv = 5.657826                                                 fv = 4.908543                                                 fv = 6.025278                                                 fv = 10.73094                                                 fv = 6.353633                                                 fv = 16.16223                                                 fv = 6.330379                                                 fv = 5.642881                                                 fv = 4.908543                                                 fv = 5.852817                                                 fv = 6.298114                                                 fv = 4.908543                                                 fv = 38.53345                                                 fv = 1.340781e+154                                                 fv = 4.908543                                                 fv = 6.025278                                                 fv = 7.409951                                                 fv = 4.908543                                                 fv = 10.38293                                                 fv = 4.112633                                                 fv = 1.340781e+154                                                 fv = 13.09092                                                 fv = 5.558411                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.21207                                                 fv = 4.590464                                                 fv = 5.872443                                                 fv = 4.112633                                                 fv = 6.886053                                                 fv = 4.112633                                                 fv = 4.46992                                                 fv = 4.112633                                                 fv = 4.112633                                                 fv = 4.112633                                                 fv = 4.112633                                                 fv = 4.998104                                                 fv = 23.9774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.948365                                                 fv = 10.38293                                                 fv = 4.577641                                                 fv = 5.986871                                                 fv = 5.678631                                                 fv = 4.859263                                                 fv = 10.38293                                                 fv = 4.911527                                                 fv = 4.869629                                                 fv = 4.372465                                                 fv = 24.55657 run # 1: f=4.112632606
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.41095                                                 fv = 30.7943                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.23518                                                 fv = 1.340781e+154                                                 fv = 16.07071                                                 fv = 23.20914                                                 fv = 131255.1                                                 fv = 1.340781e+154                                                 fv = 9.981902                                                 fv = 31.27177                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.981902                                                 fv = 1.340781e+154                                                 fv = 221.8162                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.088508                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42.36438                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 241605.9                                                 fv = 9.624315                                                 fv = 13.98577                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.12069                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.404451                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.17022                                                 fv = 24.05701                                                 fv = 14.02942                                                 fv = 383936                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5790358053                                                 fv = 383638816                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.17022                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.088508                                                 fv = 9.088508                                                 fv = 11.17022                                                 fv = 15.18356                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37739.37                                                 fv = 13.98808                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.848819                                                 fv = 1.340781e+154                                                 fv = 288006397033                                                 fv = 21059.63                                                 fv = 10.93506                                                 fv = 1.340781e+154                                                 fv = 10.93506                                                 fv = 10.93506                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.57297                                                 fv = 32.20734                                                 fv = 9.088508                                                 fv = 5120.339                                                 fv = 11.79405                                                 fv = 9.436052                                                 fv = 10.36315                                                 fv = 9.309213                                                 fv = 13.4107                                                 fv = 11.15438                                                 fv = 11.16523                                                 fv = 8.450007                                                 fv = 7.907928                                                 fv = 1.340781e+154                                                 fv = 7.299734                                                 fv = 12.64575                                                 fv = 1.340781e+154                                                 fv = 11.58896                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.299734                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.51001                                                 fv = 34.88397                                                 fv = 1.340781e+154                                                 fv = 12.04947                                                 fv = 7.299734                                                 fv = 7.484435                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.299734                                                 fv = 24.77183                                                 fv = 7.299734                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 67.38787                                                 fv = 1.340781e+154                                                 fv = 13.29026                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.04947                                                 fv = 6.840734                                                 fv = 1518.306                                                 fv = 1.340781e+154                                                 fv = 8.325275                                                 fv = 10.90579                                                 fv = 6.840734                                                 fv = 373240.8                                                 fv = 1.340781e+154                                                 fv = 107.206                                                 fv = 9.595264                                                 fv = 9.595264                                                 fv = 1.340781e+154                                                 fv = 15.84403                                                 fv = 28.61572                                                 fv = 10.90579                                                 fv = 1.340781e+154                                                 fv = 32.74816                                                 fv = 1.340781e+154                                                 fv = 8.067539                                                 fv = 10.81464                                                 fv = 1.340781e+154                                                 fv = 6.840734                                                 fv = 1.340781e+154                                                 fv = 26.7364                                                 fv = 272.6442                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.12951                                                 fv = 8.861667                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 238.9653                                                 fv = 9.434158                                                 fv = 18.55403                                                 fv = 10.05631                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 26.19209                                                 fv = 1.340781e+154                                                 fv = 18.72073                                                 fv = 1.340781e+154                                                 fv = 101.7944                                                 fv = 6.840734                                                 fv = 11765.75                                                 fv = 26.9641                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.10108                                                 fv = 1.340781e+154                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 13.6213                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 6.840734                                                 fv = 9.166866 run # 3: f=6.840733821
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 608
#> best    : 4.112632606
#> worst   : 1.340780793e+154
#> solution: x = ( 0 3 0 3 0 2 1 0 0 0 1 1 3 1 1 1 1 1 1 1 ) f(x) = 4.112632606
#> 
#>                                                 fv = 4.112633                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (5): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.361966