Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 19.22697                                                 fv = 16.11574                                                 fv = 22806871                                                 fv = 15.15351                                                 fv = 120.984                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.19383                                                 fv = 16.08919                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.19383                                                 fv = 1.340781e+154                                                 fv = 15.27071                                                 fv = 54.99952                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 49.13153                                                 fv = 37.69955                                                 fv = 3736836229                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.15351                                                 fv = 287986.4                                                 fv = 1.340781e+154                                                 fv = 10169.25                                                 fv = 22.19383                                                 fv = 17790.17                                                 fv = 22.19383                                                 fv = 50.28174                                                 fv = 17.18507                                                 fv = 133.9999                                                 fv = 39914.42                                                 fv = 950.6911                                                 fv = 1734.068                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 142.51                                                 fv = 15.15351                                                 fv = 15.15351                                                 fv = 26.9713                                                 fv = 15.15351                                                 fv = 1.340781e+154                                                 fv = 491.9696                                                 fv = 3298.981                                                 fv = 1.340781e+154                                                 fv = 9565.863                                                 fv = 67.04074                                                 fv = 16.62212                                                 fv = 24.62485                                                 fv = 17.68468                                                 fv = 18.78756                                                 fv = 36.11085                                                 fv = 37.69955                                                 fv = 16.37931                                                 fv = 135.0856                                                 fv = 187.3459                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 40.71723                                                 fv = 15.15351                                                 fv = 15.15351                                                 fv = 15.15351                                                 fv = 15.15351                                                 fv = 1.340781e+154                                                 fv = 13.87907                                                 fv = 2140.842                                                 fv = 1.340781e+154                                                 fv = 53.51672                                                 fv = 13.87907                                                 fv = 184.9953                                                 fv = 13.87907                                                 fv = 10.02032                                                 fv = 1.340781e+154                                                 fv = 14.96569                                                 fv = 14.01388                                                 fv = 10.44792                                                 fv = 8.989431                                                 fv = 20.39002                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.939598                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.789352                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.789352                                                 fv = 8.53848                                                 fv = 1.340781e+154                                                 fv = 7.939598                                                 fv = 7.939598                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.939598                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.84056                                                 fv = 8.187795                                                 fv = 1.340781e+154                                                 fv = 3479828                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.45739                                                 fv = 11.72448                                                 fv = 452.1957                                                 fv = 14.30474                                                 fv = 16.10773                                                 fv = 16.80735                                                 fv = 7.015852                                                 fv = 2920.583                                                 fv = 1.340781e+154                                                 fv = 14.77033                                                 fv = 1.340781e+154                                                 fv = 12.32921                                                 fv = 11.07361                                                 fv = 16.7264                                                 fv = 7.056498                                                 fv = 7.297454                                                 fv = 7.015852                                                 fv = 9.176012                                                 fv = 31.78303                                                 fv = 9.20893                                                 fv = 7.015852                                                 fv = 7.967655                                                 fv = 7.967655                                                 fv = 11.28431                                                 fv = 7.326879                                                 fv = 371.7839                                                 fv = 7.326879                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2967.383                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.335666                                                 fv = 7.015852                                                 fv = 9.254867                                                 fv = 9.401657                                                 fv = 8.676021                                                 fv = 7.015852                                                 fv = 14.91305                                                 fv = 9.419549                                                 fv = 96.19463                                                 fv = 7.015852                                                 fv = 12.41726                                                 fv = 6.825467                                                 fv = 9.747819                                                 fv = 1.340781e+154                                                 fv = 7.297454                                                 fv = 7.420355                                                 fv = 12.83297                                                 fv = 1.340781e+154                                                 fv = 12.83297                                                 fv = 9.729279                                                 fv = 10.71504                                                 fv = 12.64331                                                 fv = 13.11499                                                 fv = 10.84694                                                 fv = 10.84694                                                 fv = 6.825467                                                 fv = 13.70698                                                 fv = 8.728667                                                 fv = 7.595369                                                 fv = 8.287219                                                 fv = 1.340781e+154                                                 fv = 10.35464                                                 fv = 12.38741                                                 fv = 14.30474                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.71297                                                 fv = 11.41365                                                 fv = 6.825467                                                 fv = 6.466667                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.027505                                                 fv = 8.592627                                                 fv = 7.970031                                                 fv = 8.921923                                                 fv = 6.466667                                                 fv = 1.340781e+154                                                 fv = 10.1022                                                 fv = 6.620669 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 8.852136                                                 fv = 6.652412                                                 fv = 6.325325                                                 fv = 76.76623                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.573269                                                 fv = 16.09155                                                 fv = 6.325325                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.756242                                                 fv = 6.325325                                                 fv = 7.490024                                                 fv = 8.592627                                                 fv = 7.717533                                                 fv = 7.02456                                                 fv = 6.931428                                                 fv = 7.895304                                                 fv = 7.895304                                                 fv = 7.02456                                                 fv = 7.02456                                                 fv = 18.66785                                                 fv = 28.44483                                                 fv = 15.1961                                                 fv = 56.34349                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.19459                                                 fv = 6.325325                                                 fv = 6.863328                                                 fv = 6.325325                                                 fv = 7.7298                                                 fv = 6.652412                                                 fv = 8.728667                                                 fv = 10.10462                                                 fv = 8.592627                                                 fv = 7.7298                                                 fv = 26.94618                                                 fv = 9.341558                                                 fv = 11.41475                                                 fv = 7.897505                                                 fv = 9.064377                                                 fv = 10.57956                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.53111                                                 fv = 6.285486                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.285486                                                 fv = 26.86924                                                 fv = 10.54674                                                 fv = 6.285486                                                 fv = 7.897505                                                 fv = 7.006373                                                 fv = 6.285486                                                 fv = 6.285486                                                 fv = 6.285486                                                 fv = 10.06132                                                 fv = 12.86916                                                 fv = 8.80101                                                 fv = 7.970031                                                 fv = 100.958                                                 fv = 11.91262                                                 fv = 6.285486                                                 fv = 7.944287                                                 fv = 1.340781e+154                                                 fv = 25.5967                                                 fv = 12.75074                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.27436 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 7.215262                                                 fv = 1.340781e+154                                                 fv = 7.500633                                                 fv = 6.072623                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.514987                                                 fv = 1.340781e+154                                                 fv = 6.072623                                                 fv = 1.340781e+154                                                 fv = 7.868621                                                 fv = 9.184334                                                 fv = 10.20694                                                 fv = 9.655878                                                 fv = 6.072623                                                 fv = 12.47569                                                 fv = 7.086028                                                 fv = 6.072623                                                 fv = 8.07537                                                 fv = 21.07888                                                 fv = 6.325325                                                 fv = 6.072623                                                 fv = 6.072623                                                 fv = 7.235185                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 0: f=6.072622589
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 367.1318                                                 fv = 25.87405                                                 fv = 147.6672                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 128.1419                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 1.340781e+154                                                 fv = 38.723                                                 fv = 93.73976                                                 fv = 594437.8                                                 fv = 1.340781e+154                                                 fv = 14.77596                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 47.65986                                                 fv = 84.21507                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 1.340781e+154                                                 fv = 61.24277                                                 fv = 1516441                                                 fv = 1.340781e+154                                                 fv = 43.94968                                                 fv = 2032.951                                                 fv = 43.46022                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 351134635                                                 fv = 884.4353                                                 fv = 232.6038                                                 fv = 1281.73                                                 fv = 8330502                                                 fv = 1.340781e+154                                                 fv = 12.03145                                                 fv = 141.9774                                                 fv = 11.10761                                                 fv = 77393174313                                                 fv = 84.21507                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 185.241                                                 fv = 11.10761                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 171.9544                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.14297                                                 fv = 134.9194                                                 fv = 11.66292                                                 fv = 3388.558                                                 fv = 115.7201                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 543.0364                                                 fv = 11.10761                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10209964                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 11.10761                                                 fv = 19.91848                                                 fv = 1.340781e+154                                                 fv = 30.51313                                                 fv = 12.96697                                                 fv = 11.10761                                                 fv = 11.10761                                                 fv = 22.64568                                                 fv = 11.10761                                                 fv = 20.16616                                                 fv = 1.340781e+154                                                 fv = 11.10761                                                 fv = 11.10761                                                 fv = 84.21507                                                 fv = 14.77596                                                 fv = 12.86163                                                 fv = 293.4331                                                 fv = 553.6545                                                 fv = 31.32867                                                 fv = 11.10761                                                 fv = 28.11459                                                 fv = 543.0364                                                 fv = 478.882 run # 1: f=11.10761299
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.38375                                                 fv = 39.5972                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.47942                                                 fv = 12.51142                                                 fv = 2308536                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.13181                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 227215756                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6917481517                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.031473                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.13181                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.719995                                                 fv = 1.340781e+154                                                 fv = 77618.35                                                 fv = 11.49281                                                 fv = 1.340781e+154                                                 fv = 14.89642                                                 fv = 31.36847                                                 fv = 1.340781e+154                                                 fv = 47.80928                                                 fv = 15.47467                                                 fv = 1.340781e+154                                                 fv = 661.6988                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.45875                                                 fv = 1.340781e+154                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 28.51174                                                 fv = 8.031473                                                 fv = 15.11386                                                 fv = 10.1587                                                 fv = 10.97467                                                 fv = 15.47467                                                 fv = 1.340781e+154                                                 fv = 19.60399                                                 fv = 9.051562                                                 fv = 12.59781                                                 fv = 8.380435                                                 fv = 9.727092                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 8.031473                                                 fv = 6.5533                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32713.87                                                 fv = 6.5533                                                 fv = 459.8394                                                 fv = 19.00795                                                 fv = 14.37565                                                 fv = 7.296485                                                 fv = 10.31772                                                 fv = 10.31772                                                 fv = 11.81534                                                 fv = 17.58366                                                 fv = 47.02055                                                 fv = 25.96584                                                 fv = 6.5533                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.12017                                                 fv = 11.76216                                                 fv = 6.5533                                                 fv = 7.89849                                                 fv = 9.548489                                                 fv = 125.8784                                                 fv = 1.340781e+154                                                 fv = 19.54684                                                 fv = 36.71232                                                 fv = 1.340781e+154                                                 fv = 15.19285                                                 fv = 12.54079                                                 fv = 11.58522                                                 fv = 26.75929                                                 fv = 8.265626                                                 fv = 6.843366                                                 fv = 9.59986                                                 fv = 14.75666                                                 fv = 11.45774                                                 fv = 1.340781e+154                                                 fv = 6.5533                                                 fv = 6.5533                                                 fv = 6.5533                                                 fv = 7.770961                                                 fv = 7.247552                                                 fv = 6.5533                                                 fv = 24.62495                                                 fv = 133.3019                                                 fv = 1.340781e+154                                                 fv = 6.5533                                                 fv = 14.37565                                                 fv = 7.921386                                                 fv = 1.340781e+154                                                 fv = 8.265626                                                 fv = 8.944428                                                 fv = 6.5533                                                 fv = 16.11948 run # 2: f=6.553299683
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 39.5972                                                 fv = 39.5972                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 39.5972                                                 fv = 39.5972                                                 fv = 39.5972                                                 fv = 39.5972                                                 fv = 39.5972                                                 fv = 1.340781e+154                                                 fv = 39.5972                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1657.151                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.747819                                                 fv = 17.30808                                                 fv = 31.31588                                                 fv = 39.5972                                                 fv = 9.747819                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4963.48                                                 fv = 32.95734                                                 fv = 2232.216                                                 fv = 26.94618                                                 fv = 15.50901                                                 fv = 12.83297                                                 fv = 47.05265                                                 fv = 144090.9                                                 fv = 140891468980                                                 fv = 32.95734                                                 fv = 15.79635                                                 fv = 9.747819                                                 fv = 9.747819                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.67092                                                 fv = 1.340781e+154                                                 fv = 53026.67                                                 fv = 1.340781e+154                                                 fv = 30.49191                                                 fv = 158.8662                                                 fv = 15.79635                                                 fv = 1.340781e+154                                                 fv = 78.5574                                                 fv = 9.747819                                                 fv = 9.747819                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.747819                                                 fv = 65.69107                                                 fv = 13.92897                                                 fv = 1.340781e+154                                                 fv = 19.99696                                                 fv = 39.5972                                                 fv = 1.340781e+154                                                 fv = 26.67092                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.79635                                                 fv = 1.340781e+154                                                 fv = 18.07739                                                 fv = 1.340781e+154                                                 fv = 16.65613                                                 fv = 1.340781e+154                                                 fv = 30.1582                                                 fv = 102.5382                                                 fv = 1.340781e+154                                                 fv = 9.747819                                                 fv = 9.747819                                                 fv = 9.747819                                                 fv = 9.747819                                                 fv = 1.340781e+154                                                 fv = 21.33729                                                 fv = 15.50901                                                 fv = 13.05769                                                 fv = 39.5972 run # 3: f=9.747818802
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 710
#> best    : 6.072622589
#> worst   : 1.340780793e+154
#> solution: x = ( 0 1 0 0 0 1 0 0 0 1 1 2 1 2 2 2 1 3 1 3 ) f(x) = 6.072622589
#> 
#>                                                 fv = 6.072623                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 18.97512