Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 20.50016                                                 fv = 13.02898                                                 fv = 1.340781e+154                                                 fv = 15.65438                                                 fv = 1.340781e+154                                                 fv = 52.38021                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 404.7369                                                 fv = 13.02898                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 52.38021                                                 fv = 1.340781e+154                                                 fv = 17.53702                                                 fv = 17.53702                                                 fv = 535.8623                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.911011                                                 fv = 17.72116                                                 fv = 1.340781e+154                                                 fv = 10.63729                                                 fv = 26.52476                                                 fv = 8.760663                                                 fv = 1459.572                                                 fv = 10.25178                                                 fv = 1.340781e+154                                                 fv = 29.92761                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.71342                                                 fv = 1713.219                                                 fv = 9.02128                                                 fv = 10.00168                                                 fv = 1.340781e+154                                                 fv = 48772.75                                                 fv = 691.0695                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 11.3747                                                 fv = 13.26253                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.49787                                                 fv = 175.3582                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.174728                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.56764                                                 fv = 13.59465                                                 fv = 8.949185                                                 fv = 9.983472                                                 fv = 9.967444                                                 fv = 10.59719                                                 fv = 11.14555                                                 fv = 38.25419                                                 fv = 11.56936                                                 fv = 160219400                                                 fv = 11.56936                                                 fv = 119.7711                                                 fv = 1.340781e+154                                                 fv = 10.45935                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.2401                                                 fv = 11.85311                                                 fv = 89806.27                                                 fv = 8.49787                                                 fv = 7.315482                                                 fv = 7.315482                                                 fv = 11004.28                                                 fv = 14.51419                                                 fv = 22126.27                                                 fv = 9.44047                                                 fv = 6.654974                                                 fv = 1.340781e+154                                                 fv = 18.42109                                                 fv = 1.340781e+154                                                 fv = 8.075616                                                 fv = 11.71402                                                 fv = 10.18033                                                 fv = 6.654974                                                 fv = 8.95461                                                 fv = 1.340781e+154                                                 fv = 7.612908                                                 fv = 6.654974                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.242685                                                 fv = 6.654974                                                 fv = 7.052319                                                 fv = 13.82636                                                 fv = 15.83772                                                 fv = 1.340781e+154                                                 fv = 9.469474                                                 fv = 22516                                                 fv = 17.73661                                                 fv = 94458544                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 111.8845                                                 fv = 6.654974                                                 fv = 6.654974                                                 fv = 8.082826                                                 fv = 9.312049                                                 fv = 6.4388                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 9.798604                                                 fv = 1.340781e+154                                                 fv = 9.494455                                                 fv = 1.340781e+154                                                 fv = 8.654801                                                 fv = 10.63781                                                 fv = 6.4388                                                 fv = 6.889669                                                 fv = 6.487906                                                 fv = 6.730624                                                 fv = 6.592026                                                 fv = 6.4388                                                 fv = 6.4388                                                 fv = 6.4388                                                 fv = 9.096933                                                 fv = 6.419945                                                 fv = 9.260713                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.318311                                                 fv = 8.392908                                                 fv = 9.260713                                                 fv = 25.37058                                                 fv = 9.752286                                                 fv = 8.68913                                                 fv = 6.068748                                                 fv = 139.8873                                                 fv = 1.340781e+154                                                 fv = 6.135553                                                 fv = 8.542489                                                 fv = 7.153778                                                 fv = 8.778278                                                 fv = 1.340781e+154                                                 fv = 6.068748                                                 fv = 6.068748                                                 fv = 8.128298                                                 fv = 6.78028                                                 fv = 11.56483                                                 fv = 9.661497                                                 fv = 9.609439                                                 fv = 8.675133                                                 fv = 1.340781e+154                                                 fv = 6.734372                                                 fv = 7.044693                                                 fv = 6.78028                                                 fv = 9.538697                                                 fv = 8.277256                                                 fv = 8.830902                                                 fv = 10.10859                                                 fv = 11.75908                                                 fv = 19.48384                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.377224                                                 fv = 6.068748                                                 fv = 6.572033                                                 fv = 1.340781e+154                                                 fv = 6.068748                                                 fv = 6.54375                                                 fv = 1.340781e+154                                                 fv = 6.842562                                                 fv = 8.243631                                                 fv = 7.043954                                                 fv = 6.068748                                                 fv = 8.866611                                                 fv = 8.840345                                                 fv = 15.75474                                                 fv = 9.752286                                                 fv = 6.842562                                                 fv = 9.707834                                                 fv = 25.67226                                                 fv = 1.340781e+154                                                 fv = 6.80805                                                 fv = 6.369526                                                 fv = 6.068748                                                 fv = 7.68984                                                 fv = 7.781809                                                 fv = 6.572033                                                 fv = 6.712096                                                 fv = 1.340781e+154                                                 fv = 15.12513                                                 fv = 6.068748                                                 fv = 6.840506                                                 fv = 1.340781e+154                                                 fv = 6.283667                                                 fv = 8.128298                                                 fv = 9.946448                                                 fv = 6.068748                                                 fv = 30.2354 run # 0: f=6.068747623
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.595823                                                 fv = 45.0957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.09279                                                 fv = 6.564127                                                 fv = 45.0957                                                 fv = 45.0957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.05703                                                 fv = 20.43643                                                 fv = 6.326605                                                 fv = 16.5524                                                 fv = 27.90726                                                 fv = 1.340781e+154                                                 fv = 14.02595                                                 fv = 1.340781e+154                                                 fv = 6.55304                                                 fv = 1.340781e+154                                                 fv = 693.6607                                                 fv = 1.340781e+154                                                 fv = 21.82234                                                 fv = 1.340781e+154                                                 fv = 9.36413                                                 fv = 1.340781e+154                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.810221                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.72087                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.502061                                                 fv = 47.37036                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 10.94479                                                 fv = 6.326605                                                 fv = 1.340781e+154                                                 fv = 132.7542                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 10.28303                                                 fv = 1.340781e+154                                                 fv = 9.045161                                                 fv = 1.340781e+154                                                 fv = 9.045161                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.171298                                                 fv = 6.55304                                                 fv = 1.340781e+154                                                 fv = 8.360371                                                 fv = 7.984153                                                 fv = 8.884818                                                 fv = 8.490498                                                 fv = 10.44494                                                 fv = 10.88921                                                 fv = 8.487989                                                 fv = 9.265783                                                 fv = 12.38747                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 10.75346                                                 fv = 1.340781e+154                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 12.70608                                                 fv = 1.340781e+154                                                 fv = 40.19891                                                 fv = 1.340781e+154                                                 fv = 8.657031                                                 fv = 9.83686                                                 fv = 6.326605                                                 fv = 6.648159                                                 fv = 7.673256                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 6.326605                                                 fv = 9.170961                                                 fv = 6.326605                                                 fv = 8.884818                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.277275                                                 fv = 149.4275                                                 fv = 1.340781e+154                                                 fv = 12.15991                                                 fv = 8.657031                                                 fv = 6.277275                                                 fv = 6.277275                                                 fv = 1.340781e+154                                                 fv = 6.277275                                                 fv = 6.277275                                                 fv = 7.596335                                                 fv = 6.277275                                                 fv = 1.340781e+154                                                 fv = 6.277275                                                 fv = 5.909627                                                 fv = 10.8532                                                 fv = 1.340781e+154                                                 fv = 7.836538                                                 fv = 9.108318                                                 fv = 7.20959                                                 fv = 7.014161                                                 fv = 6.058881                                                 fv = 5.872219                                                 fv = 13.59837                                                 fv = 1.340781e+154                                                 fv = 6.415542                                                 fv = 1.340781e+154                                                 fv = 6.313651                                                 fv = 7.153403                                                 fv = 6.06029                                                 fv = 9.919718                                                 fv = 5.872219                                                 fv = 6.247153                                                 fv = 5.909627                                                 fv = 5.872219                                                 fv = 7.203647                                                 fv = 5.872219                                                 fv = 5.872219                                                 fv = 1.340781e+154                                                 fv = 6.295057                                                 fv = 6.158997                                                 fv = 6.388925                                                 fv = 1.340781e+154                                                 fv = 7.086547                                                 fv = 5.872219                                                 fv = 9.653827 run # 1: f=5.872218992
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.94499                                                 fv = 45.0957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33.87437                                                 fv = 59.47754                                                 fv = 7.347626                                                 fv = 871.9656                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.2864                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.59261                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31.78814                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.344371                                                 fv = 7.344371                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.529808                                                 fv = 12.07939                                                 fv = 5.107919                                                 fv = 11.97998                                                 fv = 170.1444                                                 fv = 13.28983                                                 fv = 50.18914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.285491                                                 fv = 18.31882                                                 fv = 31.33273                                                 fv = 30.03281                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.06933                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 172.3882                                                 fv = 19.19521                                                 fv = 189.499                                                 fv = 128.65                                                 fv = 1.340781e+154                                                 fv = 51.1048                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.111141                                                 fv = 1.340781e+154                                                 fv = 5.773451                                                 fv = 1.340781e+154                                                 fv = 60.63072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.113918                                                 fv = 4.113918                                                 fv = 1.340781e+154                                                 fv = 11.8289                                                 fv = 8.212815                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.74702                                                 fv = 8.846544                                                 fv = 30.92044                                                 fv = 8.212815                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.793295                                                 fv = 3.793295                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 109.0853                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29591.96                                                 fv = 3.793295                                                 fv = 1.340781e+154                                                 fv = 39.80118                                                 fv = 782.0899                                                 fv = 39.80118                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.52723                                                 fv = 1.340781e+154                                                 fv = 5.451283                                                 fv = 9.035298e+13                                                 fv = 7.455691                                                 fv = 45.53715                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.54702                                                 fv = 32.29232                                                 fv = 1.340781e+154                                                 fv = 3.793295                                                 fv = 5.77659                                                 fv = 186.1257                                                 fv = 4.470465                                                 fv = 1.340781e+154                                                 fv = 3.793295                                                 fv = 4.943014                                                 fv = 5.451283                                                 fv = 6.277818                                                 fv = 7.993625                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.793295                                                 fv = 3.793295                                                 fv = 187.6506                                                 fv = 12.93592                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.926344                                                 fv = 10.38317                                                 fv = 3.793295                                                 fv = 6.788813                                                 fv = 4.243136                                                 fv = 6.743835                                                 fv = 5.451283                                                 fv = 5.451283                                                 fv = 5.991891                                                 fv = 6.534774                                                 fv = 6.660171                                                 fv = 6.405371                                                 fv = 9.065617                                                 fv = 9.462964                                                 fv = 3.793295                                                 fv = 19.22268                                                 fv = 1.340781e+154                                                 fv = 757.8547                                                 fv = 4.618515                                                 fv = 11.91031                                                 fv = 348001.3                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.374047                                                 fv = 5.451283                                                 fv = 3.793295                                                 fv = 18.76349                                                 fv = 4.626923                                                 fv = 4.533703                                                 fv = 4.470465                                                 fv = 88.91541                                                 fv = 4.274683                                                 fv = 4.113918                                                 fv = 4.943014                                                 fv = 6.334144                                                 fv = 4.243136                                                 fv = 3.793295                                                 fv = 4.417671                                                 fv = 5.149757                                                 fv = 27.89175                                                 fv = 3.793295                                                 fv = 15.86613                                                 fv = 4.399843                                                 fv = 5.345072                                                 fv = 1.340781e+154                                                 fv = 3.504256                                                 fv = 3.504256                                                 fv = 1.340781e+154                                                 fv = 3.504256                                                 fv = 57.88403                                                 fv = 18.43117                                                 fv = 3.504256                                                 fv = 4.728078                                                 fv = 7.752639                                                 fv = 4.618836                                                 fv = 3.504256                                                 fv = 4.495385                                                 fv = 1.340781e+154                                                 fv = 3.928799                                                 fv = 3.504256                                                 fv = 3.504256                                                 fv = 4.399843                                                 fv = 1.340781e+154                                                 fv = 5.42042                                                 fv = 3.886121                                                 fv = 5.278903                                                 fv = 3.504256                                                 fv = 11.9957                                                 fv = 4.189896                                                 fv = 18.31354                                                 fv = 1.340781e+154                                                 fv = 5.992675                                                 fv = 11.59968                                                 fv = 3.504256                                                 fv = 3.928799                                                 fv = 3.793295                                                 fv = 4.040676                                                 fv = 10.22622                                                 fv = 3.782903                                                 fv = 3.944285                                                 fv = 16.10301                                                 fv = 4.036393                                                 fv = 3.504256                                                 fv = 3.504256                                                 fv = 4.170615                                                 fv = 3.782903                                                 fv = 5.181028                                                 fv = 3.504256                                                 fv = 4.277588                                                 fv = 3.504256                                                 fv = 3.896173                                                 fv = 8.820501 run # 2: f=3.504255954
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.66405                                                 fv = 45.0957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.66405                                                 fv = 10.41568                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.77391                                                 fv = 7.632032                                                 fv = 41726708                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 196006                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7333.404                                                 fv = 25.42872                                                 fv = 1.340781e+154                                                 fv = 11.66405                                                 fv = 2078.234                                                 fv = 10.10725                                                 fv = 79.56037                                                 fv = 7.62119                                                 fv = 1739.56                                                 fv = 15.66904                                                 fv = 23.45049                                                 fv = 10.82386                                                 fv = 8191192                                                 fv = 26.21114                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 209.3922                                                 fv = 1.340781e+154                                                 fv = 9.520237                                                 fv = 7.62119                                                 fv = 40112748                                                 fv = 7.069217                                                 fv = 7.069217                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.247976                                                 fv = 41.25656                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 98552.71                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.363328                                                 fv = 1.340781e+154                                                 fv = 6.247976                                                 fv = 209.3922                                                 fv = 1.340781e+154                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 142.1379                                                 fv = 6.247976                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 84.37778                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 40449.55                                                 fv = 6.247976                                                 fv = 6.630916                                                 fv = 7.210414                                                 fv = 6.247976                                                 fv = 7.210414                                                 fv = 7.210414                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 7.069217                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.553139                                                 fv = 9.146254                                                 fv = 1.340781e+154                                                 fv = 6.30067                                                 fv = 1.340781e+154                                                 fv = 9.232095                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.20912                                                 fv = 8.42967                                                 fv = 6.247976                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.461245                                                 fv = 6151.716                                                 fv = 163.843                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 115.594                                                 fv = 6.247976                                                 fv = 8.338282                                                 fv = 6.247976                                                 fv = 1.340781e+154                                                 fv = 11.6233                                                 fv = 6.792251                                                 fv = 39.21497                                                 fv = 8.683101                                                 fv = 1.340781e+154                                                 fv = 20.56852                                                 fv = 8.143291                                                 fv = 6.247976                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.760883                                                 fv = 211.7882                                                 fv = 7.210414                                                 fv = 9.181329                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 6.247976                                                 fv = 7.677067                                                 fv = 6.247976                                                 fv = 6.982783                                                 fv = 7.069217                                                 fv = 9.146254                                                 fv = 9.146254                                                 fv = 39.21497                                                 fv = 7.300768 run # 4: f=6.247975731
#> 
#> bb eval : 801
#> best    : 3.504255954
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 3 0 1 1 1 0 0 1 1 1 1 6 1 1 2 2 1 ) f(x) = 3.504255954
#> 
#>                                                 fv = 3.504256                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...          

# Score the predictions
predictions$score()
#> regr.mse 
#>  6.36385