Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 19.95417                                                 fv = 10.3519                                                 fv = 1.340781e+154                                                 fv = 10.58545                                                 fv = 1.340781e+154                                                 fv = 60.27438                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 88.59958                                                 fv = 10.3519                                                 fv = 1.340781e+154                                                 fv = 6506373                                                 fv = 60.27438                                                 fv = 1.340781e+154                                                 fv = 36.72076                                                 fv = 36.72076                                                 fv = 35.62035                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.056463                                                 fv = 12.60649                                                 fv = 1.340781e+154                                                 fv = 9.04745                                                 fv = 30.85198                                                 fv = 7336479014                                                 fv = 416761313                                                 fv = 1.340781e+154                                                 fv = 503.0861                                                 fv = 12.67383                                                 fv = 1.340781e+154                                                 fv = 26700.15                                                 fv = 769.5505                                                 fv = 194.1071                                                 fv = 11.65314                                                 fv = 12.8002                                                 fv = 11.45251                                                 fv = 793.1789                                                 fv = 15.37552                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 72533.84                                                 fv = 1.340781e+154                                                 fv = 456.2756                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.22289                                                 fv = 6529.241                                                 fv = 15.54356                                                 fv = 18.43843                                                 fv = 1.340781e+154                                                 fv = 6.889044                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 196.6633                                                 fv = 10.15428                                                 fv = 718.1496                                                 fv = 7247.961                                                 fv = 37741.91                                                 fv = 13.04113                                                 fv = 1.340781e+154                                                 fv = 45.30859                                                 fv = 12.72176                                                 fv = 56.78667                                                 fv = 4019.838                                                 fv = 83.84264                                                 fv = 6.889044                                                 fv = 9.838055                                                 fv = 9.504499                                                 fv = 6.889044                                                 fv = 12.07773                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.2901                                                 fv = 8.512511                                                 fv = 7.171063                                                 fv = 9.485544                                                 fv = 19.4209                                                 fv = 8.745537                                                 fv = 1.340781e+154                                                 fv = 6.889044                                                 fv = 6.889044                                                 fv = 7.508325                                                 fv = 1.340781e+154                                                 fv = 12.29334                                                 fv = 6.889044                                                 fv = 7.32364                                                 fv = 6.97914                                                 fv = 6.889044                                                 fv = 8.325772                                                 fv = 12.68193                                                 fv = 11.97009                                                 fv = 10.72656                                                 fv = 1.340781e+154                                                 fv = 21.09827                                                 fv = 5.972119                                                 fv = 1.340781e+154                                                 fv = 16.65794                                                 fv = 6.432778                                                 fv = 6.223805                                                 fv = 6.909262                                                 fv = 6.507188                                                 fv = 5.972119                                                 fv = 5.972119                                                 fv = 6.297685                                                 fv = 8.3357                                                 fv = 6.355555                                                 fv = 5.972119                                                 fv = 5.972119                                                 fv = 5.972119                                                 fv = 1.340781e+154                                                 fv = 6.647436                                                 fv = 7.730447                                                 fv = 5.972119                                                 fv = 8.19624                                                 fv = 6.669952                                                 fv = 5.972119                                                 fv = 5.972119                                                 fv = 6.192939                                                 fv = 5.972119                                                 fv = 6.057144                                                 fv = 9.594689                                                 fv = 17.2275 run # 0: f=5.972118699
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.693088                                                 fv = 45.7973                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.693088                                                 fv = 23.35068                                                 fv = 2017.981                                                 fv = 7.990981                                                 fv = 369.2262                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.4567                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.44297                                                 fv = 1.340781e+154                                                 fv = 44796.64                                                 fv = 9.051654                                                 fv = 1.340781e+154                                                 fv = 15.51251                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.990981                                                 fv = 50.75554                                                 fv = 7.990981                                                 fv = 8.526756                                                 fv = 5.984959                                                 fv = 24.58612                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.675308                                                 fv = 6.662833                                                 fv = 5.984959                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.940499                                                 fv = 1.340781e+154                                                 fv = 13.18239                                                 fv = 7.683648                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.863557                                                 fv = 10.11018                                                 fv = 1.340781e+154                                                 fv = 5.863557                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.338311                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.254152                                                 fv = 9.24385                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.62268                                                 fv = 5.863557                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.60967                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.685812                                                 fv = 7.367428                                                 fv = 7.367428                                                 fv = 5.863557                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 13.13168                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2025.635                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.243557                                                 fv = 36.9594                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.237115                                                 fv = 1.340781e+154                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 10.41685                                                 fv = 13.25758                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 112.8368                                                 fv = 13.13168                                                 fv = 723.9614                                                 fv = 6.79773                                                 fv = 27.9467                                                 fv = 5.926974                                                 fv = 5.956356                                                 fv = 11.9177                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 23.97035                                                 fv = 5.799975                                                 fv = 5.799975                                                 fv = 7.153359                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 6.450961                                                 fv = 9.990758                                                 fv = 5.863557                                                 fv = 18.0258                                                 fv = 7468.624                                                 fv = 694.9172                                                 fv = 43.55813                                                 fv = 5.799975                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.165497                                                 fv = 5.593424                                                 fv = 145.1203                                                 fv = 1.340781e+154                                                 fv = 15075.19                                                 fv = 1.340781e+154                                                 fv = 5.593424                                                 fv = 7.004946                                                 fv = 8.215223                                                 fv = 5.593424                                                 fv = 5.593424                                                 fv = 8.757651                                                 fv = 9.219596                                                 fv = 5.593424                                                 fv = 5.593424                                                 fv = 5.945082                                                 fv = 7.311037                                                 fv = 6.995598                                                 fv = 7.299758                                                 fv = 15.60596                                                 fv = 63.3953                                                 fv = 50.47277                                                 fv = 36.46407                                                 fv = 5.593424                                                 fv = 38.41018                                                 fv = 6.744888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=5.593424057
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9381.787                                                 fv = 45.7973                                                 fv = 21.267                                                 fv = 14.64305                                                 fv = 14.64305                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.64305                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.431386                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 993580054840 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1030415412                                                 fv = 41113919888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.028213                                                 fv = 7.750483                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 809.0423                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 268.1657                                                 fv = 1.340781e+154                                                 fv = 30.30824                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 993580054840                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.40469                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.028213                                                 fv = 7.028213                                                 fv = 1.340781e+154                                                 fv = 8.844403                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.91033                                                 fv = 6.959981                                                 fv = 1.340781e+154                                                 fv = 8.844403                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.91033                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 4742.186                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 6.341782                                                 fv = 6.341782                                                 fv = 10.30533                                                 fv = 8.944237                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.844403                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 6.931768                                                 fv = 1760.439                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 7.415589                                                 fv = 17.29885                                                 fv = 6.91033                                                 fv = 1.340781e+154                                                 fv = 6.892308                                                 fv = 8.358676                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 38680.04                                                 fv = 1.340781e+154                                                 fv = 38.02984                                                 fv = 10.30533                                                 fv = 6.83018                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.469276                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 6.438096                                                 fv = 6.70637                                                 fv = 10.01517                                                 fv = 10.16316                                                 fv = 10.30533                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 6.306145                                                 fv = 6.385523                                                 fv = 6.158717                                                 fv = 419612.5                                                 fv = 1.340781e+154                                                 fv = 10.30533                                                 fv = 6.158717                                                 fv = 6.158717                                                 fv = 6.357587                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.158717                                                 fv = 7.032668                                                 fv = 6.158717                                                 fv = 1.340781e+154                                                 fv = 38.21939                                                 fv = 1.340781e+154                                                 fv = 6.158717                                                 fv = 6.158717                                                 fv = 6.158717                                                 fv = 24.91976                                                 fv = 7.43028                                                 fv = 7.181028                                                 fv = 6.306145                                                 fv = 7.497439                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.158717                                                 fv = 24.71762                                                 fv = 12.55985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.158717                                                 fv = 7.400862                                                 fv = 6.158717                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2538.947                                                 fv = 10.30533 run # 3: f=6.158716922
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 595
#> best    : 5.593424057
#> worst   : 1.340780793e+154
#> solution: x = ( 0 4 1 2 0 0 1 0 0 0 1 1 1 1 1 1 1 1 1 1 ) f(x) = 5.593424057
#> 
#>                                                 fv = 5.593424                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 211.5215