Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 12.70948                                                 fv = 13.76899                                                 fv = 6.37686                                                 fv = 277028.9                                                 fv = 1.340781e+154                                                 fv = 7.231071                                                 fv = 8.306735                                                 fv = 30666418                                                 fv = 1.340781e+154                                                 fv = 40.50278                                                 fv = 20.16709                                                 fv = 1.340781e+154                                                 fv = 38.14143                                                 fv = 1.340781e+154                                                 fv = 90.82831                                                 fv = 1.340781e+154                                                 fv = 8.306735                                                 fv = 8.306735                                                 fv = 1.340781e+154                                                 fv = 70.34557                                                 fv = 1.340781e+154                                                 fv = 18.87164                                                 fv = 6316359                                                 fv = 7631.708                                                 fv = 12.7007                                                 fv = 14.76077                                                 fv = 10.55749                                                 fv = 6.37686                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.37686                                                 fv = 6.37686                                                 fv = 1.340781e+154                                                 fv = 22.67992                                                 fv = 42.10737                                                 fv = 120.7513                                                 fv = 678.0316                                                 fv = 149.4855                                                 fv = 1.340781e+154                                                 fv = 20570.3                                                 fv = 13.39639                                                 fv = 251.7264                                                 fv = 8.860215                                                 fv = 1.340781e+154                                                 fv = 27.28849                                                 fv = 12.31572                                                 fv = 6.37686                                                 fv = 8.892169                                                 fv = 8.036333                                                 fv = 8.566824                                                 fv = 1.340781e+154                                                 fv = 6.37686                                                 fv = 1.340781e+154                                                 fv = 6.37686                                                 fv = 6.37686                                                 fv = 6.37686                                                 fv = 1.340781e+154                                                 fv = 85.84727                                                 fv = 9.126906                                                 fv = 7.261638                                                 fv = 7.154475 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 6.945517                                                 fv = 6.911053                                                 fv = 4.926922                                                 fv = 28.06745                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.231071                                                 fv = 24.17859                                                 fv = 24.17859                                                 fv = 4.423809                                                 fv = 2785194                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.854973                                                 fv = 24.90205                                                 fv = 4.423809                                                 fv = 4.423809                                                 fv = 4.423809                                                 fv = 4.423809                                                 fv = 4.423809                                                 fv = 5.507397                                                 fv = 5.294169                                                 fv = 4.926922                                                 fv = 6.319332                                                 fv = 6.069761                                                 fv = 4.423809                                                 fv = 5.761284                                                 fv = 74.9401                                                 fv = 24.4861                                                 fv = 579.5033                                                 fv = 161.9575                                                 fv = 15.92212                                                 fv = 10039.03                                                 fv = 7.196789                                                 fv = 1.340781e+154                                                 fv = 4.219208                                                 fv = 15.54105                                                 fv = 1.340781e+154                                                 fv = 6.709582                                                 fv = 7.837211                                                 fv = 6.437468                                                 fv = 29.47726                                                 fv = 5.371183                                                 fv = 5.371183                                                 fv = 5.371183                                                 fv = 9.935071                                                 fv = 439.1584                                                 fv = 5.007354                                                 fv = 4.884668                                                 fv = 1.340781e+154                                                 fv = 363.9343                                                 fv = 4.219208                                                 fv = 4.219208                                                 fv = 4.219208                                                 fv = 4.028982                                                 fv = 1.340781e+154                                                 fv = 5.63145                                                 fv = 74.9401                                                 fv = 12.27803                                                 fv = 1.340781e+154                                                 fv = 5.776315                                                 fv = 1.340781e+154                                                 fv = 4.028982                                                 fv = 5.621839                                                 fv = 4.028982                                                 fv = 4.028982                                                 fv = 4.526208                                                 fv = 4.028982                                                 fv = 34.08016                                                 fv = 5.409494                                                 fv = 4.028982                                                 fv = 1.340781e+154                                                 fv = 4.028982                                                 fv = 20.6046                                                 fv = 142.574                                                 fv = 24.46411                                                 fv = 5.475661                                                 fv = 111.8398                                                 fv = 24.91497                                                 fv = 49908.5                                                 fv = 1.340781e+154                                                 fv = 5.079357                                                 fv = 54366.56                                                 fv = 5.238178                                                 fv = 11.83349                                                 fv = 10487.38                                                 fv = 8.480076                                                 fv = 15.64432                                                 fv = 15.64432                                                 fv = 1.340781e+154                                                 fv = 38263.7                                                 fv = 6.188305                                                 fv = 5.607523                                                 fv = 5.621839                                                 fv = 4.028982                                                 fv = 4.028982                                                 fv = 4.028982                                                 fv = 4.028982                                                 fv = 4.564146                                                 fv = 4.756931                                                 fv = 4.278065                                                 fv = 4.028982                                                 fv = 12.00746 run # 0: f=4.028982289
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.30322                                                 fv = 28.06745                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.30735                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.575105                                                 fv = 8.442501                                                 fv = 8.442501                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.442501                                                 fv = 8.794505                                                 fv = 10.27141                                                 fv = 1.340781e+154                                                 fv = 211802.2                                                 fv = 60.60635                                                 fv = 14.2486                                                 fv = 272563                                                 fv = 17.22813                                                 fv = 28.06745                                                 fv = 175.7764                                                 fv = 175.7764                                                 fv = 1.340781e+154                                                 fv = 8.794505                                                 fv = 28735.99                                                 fv = 215734.8                                                 fv = 12.52686                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.28249                                                 fv = 1.340781e+154                                                 fv = 8.442501                                                 fv = 9.01371                                                 fv = 1.340781e+154                                                 fv = 12.79531                                                 fv = 8.442501                                                 fv = 1.340781e+154                                                 fv = 424.297                                                 fv = 8.442501                                                 fv = 1.340781e+154                                                 fv = 20.41468                                                 fv = 426.677                                                 fv = 8.376234                                                 fv = 1.340781e+154                                                 fv = 22.59249                                                 fv = 1.340781e+154                                                 fv = 8.376234                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.22484                                                 fv = 11.22484                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.376234                                                 fv = 1.340781e+154                                                 fv = 1482.153                                                 fv = 10.99155                                                 fv = 8.376234                                                 fv = 9.092515                                                 fv = 20.59035                                                 fv = 9.146241                                                 fv = 1.340781e+154                                                 fv = 1028.573                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.83095                                                 fv = 19.68221                                                 fv = 23.67219                                                 fv = 11.37673                                                 fv = 19.98647                                                 fv = 24.73166                                                 fv = 49.05239                                                 fv = 10.60046                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.376234                                                 fv = 8.376234                                                 fv = 8.376234                                                 fv = 7.824331                                                 fv = 9.572884                                                 fv = 1.340781e+154                                                 fv = 52.08473                                                 fv = 46.58438                                                 fv = 5.7936                                                 fv = 20.59035                                                 fv = 1.340781e+154                                                 fv = 11.47191                                                 fv = 20.59035                                                 fv = 1.340781e+154                                                 fv = 8.199746                                                 fv = 5.603545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.603545                                                 fv = 5.603545                                                 fv = 5.7703                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.603545                                                 fv = 5.7936                                                 fv = 5.7936                                                 fv = 8.199746                                                 fv = 5.603545                                                 fv = 5.603545                                                 fv = 5.603545                                                 fv = 4.869464                                                 fv = 4.869464                                                 fv = 1.340781e+154                                                 fv = 4.497346                                                 fv = 3.758062                                                 fv = 1.340781e+154                                                 fv = 4.571986                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 24.27997                                                 fv = 6.514304                                                 fv = 7.37495                                                 fv = 8.748335                                                 fv = 75.06854                                                 fv = 211558.1                                                 fv = 5.102301                                                 fv = 4.725864                                                 fv = 4.044655                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36499856                                                 fv = 143.7756                                                 fv = 12.91753                                                 fv = 14.69131                                                 fv = 5244001037                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.05783                                                 fv = 446.3923                                                 fv = 1.340781e+154                                                 fv = 8.94127                                                 fv = 4.80314                                                 fv = 8.94127                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 8.367107                                                 fv = 4.718974                                                 fv = 3.758062                                                 fv = 8.94127                                                 fv = 3.758062                                                 fv = 6.625851                                                 fv = 47348.75                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.06745                                                 fv = 1.340781e+154                                                 fv = 5.678121                                                 fv = 3875.772                                                 fv = 3.758062                                                 fv = 4.479731                                                 fv = 5.295631                                                 fv = 5.540249                                                 fv = 4.120574                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 4.589847                                                 fv = 1.340781e+154                                                 fv = 3.758062                                                 fv = 4.406578                                                 fv = 4.622283                                                 fv = 6.83888                                                 fv = 4.917539                                                 fv = 1.340781e+154                                                 fv = 4.917539                                                 fv = 3.758062                                                 fv = 10.59039                                                 fv = 1.340781e+154                                                 fv = 10.19689                                                 fv = 4.725864                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 4.84154                                                 fv = 4.756173                                                 fv = 1.340781e+154                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 4.044655                                                 fv = 4.044655                                                 fv = 4.725864                                                 fv = 3.957353                                                 fv = 8.367107                                                 fv = 19.88429                                                 fv = 9.739897                                                 fv = 1.340781e+154                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 1.340781e+154                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 3.758062                                                 fv = 4.30879                                                 fv = 3.758062                                                 fv = 4.718974 run # 1: f=3.758061909
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 114.6582                                                 fv = 28.06745                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 426902445                                                 fv = 9.656673                                                 fv = 16562381                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 28.06745                                                 fv = 4.850791                                                 fv = 9.394464                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 1828349505                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 4452.574                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 380124.3                                                 fv = 57653.77                                                 fv = 1.340781e+154                                                 fv = 4.430193                                                 fv = 1.340781e+154                                                 fv = 3664490                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.725239                                                 fv = 219362885                                                 fv = 4.430193                                                 fv = 4.314018                                                 fv = 4.430193                                                 fv = 4.427057                                                 fv = 4.132659                                                 fv = 5.295631                                                 fv = 4.430193                                                 fv = 30.91508                                                 fv = 219361.9                                                 fv = 330590.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.579854                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.556086                                                 fv = 5.622113                                                 fv = 4.132659                                                 fv = 5.827825                                                 fv = 5.622113                                                 fv = 5.622113                                                 fv = 4.774625                                                 fv = 7.695668                                                 fv = 5.939181                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 2038.253                                                 fv = 393.9628                                                 fv = 1.340781e+154                                                 fv = 13.77973                                                 fv = 1.340781e+154                                                 fv = 9.394464                                                 fv = 54998935759                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.579854                                                 fv = 9552.747                                                 fv = 4.132659                                                 fv = 4.132659                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 4.132659                                                 fv = 7.116735                                                 fv = 26.15506                                                 fv = 49293.08                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 4.364214                                                 fv = 3727.181                                                 fv = 10.24797                                                 fv = 31.67519                                                 fv = 1.340781e+154                                                 fv = 7.859046                                                 fv = 14.47065                                                 fv = 4.868172                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.152853                                                 fv = 4.132659                                                 fv = 4.132659                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 4.701604                                                 fv = 7.496453                                                 fv = 4.645321                                                 fv = 6.247113                                                 fv = 7.002922                                                 fv = 4.824755                                                 fv = 4.132659                                                 fv = 4.132659                                                 fv = 1.340781e+154                                                 fv = 4.132659                                                 fv = 4.132659                                                 fv = 8.062552                                                 fv = 4.390464                                                 fv = 4.132659                                                 fv = 4.345298 run # 3: f=4.132659051
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.746073                                                 fv = 28.06745                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.7651                                                 fv = 47.32559                                                 fv = 251.3432                                                 fv = 38.83537                                                 fv = 1.340781e+154                                                 fv = 27.68238                                                 fv = 319.0848                                                 fv = 17.69982                                                 fv = 18.67561                                                 fv = 1.340781e+154                                                 fv = 34.08359                                                 fv = 2541.72                                                 fv = 1.340781e+154                                                 fv = 1357.05                                                 fv = 8.746073                                                 fv = 8.746073                                                 fv = 708680.9                                                 fv = 8.746073                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.3666                                                 fv = 14.04874                                                 fv = 155.8234                                                 fv = 1.340781e+154                                                 fv = 253328939                                                 fv = 1.340781e+154                                                 fv = 5.43087                                                 fv = 344.3356                                                 fv = 1.340781e+154                                                 fv = 20.92349                                                 fv = 5.43087                                                 fv = 69983261                                                 fv = 5.43087                                                 fv = 1.340781e+154                                                 fv = 12.79704                                                 fv = 1.340781e+154                                                 fv = 7.431048                                                 fv = 6.028911                                                 fv = 1.340781e+154                                                 fv = 5.43087                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 109.1682                                                 fv = 1.340781e+154                                                 fv = 6.007107                                                 fv = 6.007107                                                 fv = 5.43087                                                 fv = 87.29463                                                 fv = 1.340781e+154                                                 fv = 414566.9                                                 fv = 1.340781e+154                                                 fv = 4.899318                                                 fv = 4.162503                                                 fv = 1.340781e+154                                                 fv = 7.717318                                                 fv = 9.313446                                                 fv = 9.788818                                                 fv = 8.860215                                                 fv = 55.61704                                                 fv = 8.180422                                                 fv = 8.860215                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.986031                                                 fv = 4.345298                                                 fv = 756442758                                                 fv = 4.345298                                                 fv = 8849505                                                 fv = 4.986031                                                 fv = 4.986031                                                 fv = 4.345298                                                 fv = 1.340781e+154                                                 fv = 4.986031                                                 fv = 1584.468                                                 fv = 1.340781e+154                                                 fv = 8.860215                                                 fv = 26.26235                                                 fv = 4.345298                                                 fv = 4.394115                                                 fv = 13065.49                                                 fv = 4.394115                                                 fv = 4.591753                                                 fv = 4.633336                                                 fv = 4.919212                                                 fv = 4.162503                                                 fv = 4.162503                                                 fv = 4.162503                                                 fv = 4.162503                                                 fv = 73.137                                                 fv = 1.340781e+154                                                 fv = 4.657792                                                 fv = 6.61009                                                 fv = 8.180422                                                 fv = 86.94131                                                 fv = 2771.869                                                 fv = 28.06745                                                 fv = 39.7959                                                 fv = 9.394464                                                 fv = 4.149191                                                 fv = 3470.185                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.149191                                                 fv = 4.986031                                                 fv = 3.999337                                                 fv = 4.118145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.898524                                                 fv = 3.999337                                                 fv = 3.999337                                                 fv = 3.999337                                                 fv = 1.340781e+154                                                 fv = 3.999337                                                 fv = 3.999337                                                 fv = 1.340781e+154                                                 fv = 3.999337                                                 fv = 3.999337                                                 fv = 4.118145                                                 fv = 4.325453                                                 fv = 4.364214                                                 fv = 1.340781e+154                                                 fv = 4.582285                                                 fv = 3.999337                                                 fv = 8.975861                                                 fv = 8.277009                                                 fv = 17.43734                                                 fv = 4.947283                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.302793                                                 fv = 1.340781e+154                                                 fv = 4.607114                                                 fv = 4.535839                                                 fv = 4.118145                                                 fv = 3.999337                                                 fv = 3.999337                                                 fv = 4.118145                                                 fv = 4.582285                                                 fv = 70.35693                                                 fv = 26.90013                                                 fv = 4.899318 run # 4: f=3.999337301
#> 
#> bb eval : 759
#> best    : 3.758061909
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 0 0 0 1 0 0 3 1 2 3 1 1 7 1 7 1 1 ) f(x) = 3.758061909
#> 
#>                                                 fv = 3.758062                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.159576