Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 24.62212                                                 fv = 14.71812                                                 fv = 8240329                                                 fv = 18.06418                                                 fv = 1.340781e+154                                                 fv = 36.99922                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 61.63064                                                 fv = 14.71812                                                 fv = 1.340781e+154                                                 fv = 33718.41                                                 fv = 36.99922                                                 fv = 1.340781e+154                                                 fv = 24.34411                                                 fv = 24.34411                                                 fv = 219.7291                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.80197                                                 fv = 21.76559                                                 fv = 1.340781e+154                                                 fv = 12.11894                                                 fv = 20.28899                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 8.721418                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.836675                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 1.340781e+154                                                 fv = 9.047493                                                 fv = 8.589361                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10605444068                                                 fv = 8.589361                                                 fv = 22.44768                                                 fv = 8.864234                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 8.589361                                                 fv = 8.589361                                                 fv = 13.71226                                                 fv = 9.683681                                                 fv = 46.02953                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.578417                                                 fv = 10.38258                                                 fv = 1.340781e+154                                                 fv = 9.785526                                                 fv = 1.340781e+154                                                 fv = 9.953823                                                 fv = 1.340781e+154                                                 fv = 10.27825                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.2558                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.52625                                                 fv = 8.578417                                                 fv = 10.95932                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44425.12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.711502                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.711502                                                 fv = 1.340781e+154                                                 fv = 9.117817                                                 fv = 9.437186                                                 fv = 7.711502                                                 fv = 1.340781e+154                                                 fv = 10.60587                                                 fv = 244058.9                                                 fv = 1.340781e+154                                                 fv = 22.78544                                                 fv = 10.22807                                                 fv = 7.854198                                                 fv = 11.68656                                                 fv = 7.854198                                                 fv = 7.854198                                                 fv = 1.340781e+154                                                 fv = 8.296029                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 47.00301                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 8.864234                                                 fv = 9.924075                                                 fv = 9.924075                                                 fv = 10.04797                                                 fv = 9.924075                                                 fv = 7.711502                                                 fv = 11.16373                                                 fv = 411.1503                                                 fv = 8.589361                                                 fv = 9.924075                                                 fv = 9.437186                                                 fv = 7.711502                                                 fv = 10.22807                                                 fv = 96.86554                                                 fv = 1.340781e+154                                                 fv = 10.76794                                                 fv = 1.340781e+154                                                 fv = 11.65589                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 11.01976                                                 fv = 7.711502                                                 fv = 8.270412                                                 fv = 7.333591                                                 fv = 17.41209                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.333591                                                 fv = 7.333591                                                 fv = 7.333591                                                 fv = 7.333591                                                 fv = 7.333591                                                 fv = 7.081894                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.589361                                                 fv = 67.1426                                                 fv = 1.340781e+154                                                 fv = 7.081894                                                 fv = 7.081894                                                 fv = 1.340781e+154                                                 fv = 7.081894                                                 fv = 1.340781e+154                                                 fv = 7.980574                                                 fv = 5.382039                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 6.280047                                                 fv = 5.382039                                                 fv = 6.598075                                                 fv = 1.340781e+154                                                 fv = 10.32588                                                 fv = 5.53078                                                 fv = 13.38632                                                 fv = 13.9619                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.421171                                                 fv = 10.67017                                                 fv = 11.24274                                                 fv = 7.96317                                                 fv = 12.35459                                                 fv = 6.741137                                                 fv = 12.77251                                                 fv = 30.9563                                                 fv = 1.340781e+154                                                 fv = 9.824217                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.382039                                                 fv = 5.382039                                                 fv = 6.716805                                                 fv = 6.142683                                                 fv = 1.340781e+154                                                 fv = 6.443032                                                 fv = 1.340781e+154                                                 fv = 6.589513                                                 fv = 6.598075                                                 fv = 7.369356                                                 fv = 47.0309 run # 0: f=5.382039224
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5224.516                                                 fv = 7028.332                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1867.406                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27685.04                                                 fv = 5101.911                                                 fv = 32665.69                                                 fv = 1.340781e+154                                                 fv = 90753217954                                                 fv = 2008.46                                                 fv = 5.242668e+12                                                 fv = 26487.28                                                 fv = 137749.6                                                 fv = 9524.052                                                 fv = 43150.05                                                 fv = 1.340781e+154                                                 fv = 9524.052                                                 fv = 9524.052                                                 fv = 302.6826                                                 fv = 19.86358                                                 fv = 47.0309                                                 fv = 11373.13                                                 fv = 19.86358                                                 fv = 10.86151                                                 fv = 588584.4                                                 fv = 1.340781e+154                                                 fv = 5810.358                                                 fv = 13485635436                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 217768.3                                                 fv = 10.18044                                                 fv = 6.776891                                                 fv = 730712.6                                                 fv = 1.340781e+154                                                 fv = 16668.75                                                 fv = 47.0309                                                 fv = 5.162609                                                 fv = 38269045                                                 fv = 1.340781e+154                                                 fv = 21.67828                                                 fv = 172.8212                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 172.8212                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11789391                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.99168                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1813.388                                                 fv = 12.41033                                                 fv = 100.7395                                                 fv = 30.08379                                                 fv = 1.340781e+154                                                 fv = 6.228384                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29761.35                                                 fv = 48.09681                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 91.57912                                                 fv = 159.3475                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.71653                                                 fv = 5.162609                                                 fv = 4.430226                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15739648032                                                 fv = 113.4009                                                 fv = 1.340781e+154                                                 fv = 435.7663                                                 fv = 249.7265                                                 fv = 1.340781e+154                                                 fv = 75940.09                                                 fv = 15.33404                                                 fv = 186.2409                                                 fv = 255.1896                                                 fv = 1.340781e+154                                                 fv = 274.418                                                 fv = 173.5089                                                 fv = 173.5089                                                 fv = 254.9208                                                 fv = 4.430226                                                 fv = 4.588479                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.430226                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.430226                                                 fv = 20.25839                                                 fv = 12.79865                                                 fv = 1.340781e+154                                                 fv = 13.52214                                                 fv = 4.430226                                                 fv = 1.340781e+154                                                 fv = 4.430226                                                 fv = 5548461                                                 fv = 13622.86                                                 fv = 4.430226                                                 fv = 155056                                                 fv = 1.340781e+154                                                 fv = 14.12084                                                 fv = 8.317026                                                 fv = 12.79865                                                 fv = 22.61325                                                 fv = 4.430226                                                 fv = 24.00069                                                 fv = 1.340781e+154                                                 fv = 13.62846                                                 fv = 5.606904                                                 fv = 4.975014                                                 fv = 6.180349                                                 fv = 173.5089                                                 fv = 4.373159                                                 fv = 13.97824                                                 fv = 1.340781e+154                                                 fv = 352.5073                                                 fv = 176.0044                                                 fv = 9.814767                                                 fv = 26.52611                                                 fv = 3.958256                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 527.7866                                                 fv = 6.091042                                                 fv = 5.281035                                                 fv = 3.958256                                                 fv = 1.340781e+154                                                 fv = 6.497491                                                 fv = 1.340781e+154                                                 fv = 4.373159                                                 fv = 4.373159                                                 fv = 8.611563                                                 fv = 4.145513                                                 fv = 6.770829                                                 fv = 6.300407                                                 fv = 1.340781e+154                                                 fv = 4.373159                                                 fv = 3.958256                                                 fv = 82440539520                                                 fv = 1.340781e+154                                                 fv = 211.6053                                                 fv = 5.606904                                                 fv = 3.7435                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.22178                                                 fv = 3.7435                                                 fv = 1.340781e+154                                                 fv = 3.7435                                                 fv = 4.344258                                                 fv = 1.340781e+154                                                 fv = 3.7435                                                 fv = 11.74328                                                 fv = 9.460068                                                 fv = 4.605151                                                 fv = 1.340781e+154                                                 fv = 7.578194                                                 fv = 7.899189                                                 fv = 17.378                                                 fv = 3.7435                                                 fv = 3.7435                                                 fv = 21.56471                                                 fv = 9148192309                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.000242                                                 fv = 3.7435                                                 fv = 3.7435                                                 fv = 5.956344                                                 fv = 4.771901                                                 fv = 4.914242                                                 fv = 1.340781e+154                                                 fv = 6.564166                                                 fv = 1.340781e+154                                                 fv = 24.38283                                                 fv = 1.340781e+154                                                 fv = 9.947274                                                 fv = 4.758244                                                 fv = 3.7435                                                 fv = 6.40613                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.424243                                                 fv = 94.91429 run # 1: f=3.743500413
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2115.532                                                 fv = 47.0309                                                 fv = 47.0309                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 47.0309                                                 fv = 1.340781e+154                                                 fv = 47.0309                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 47.0309                                                 fv = 47.0309                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1258596                                                 fv = 1.340781e+154                                                 fv = 9.223922                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 54587978 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 22226.32                                                 fv = 7.35601                                                 fv = 13.14195                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.16885                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4978911                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.88158                                                 fv = 1451669                                                 fv = 7.35601                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.967116e+13                                                 fv = 39.54457                                                 fv = 1.340781e+154                                                 fv = 83071.27                                                 fv = 422972664                                                 fv = 286.7903                                                 fv = 8270.75                                                 fv = 1.340781e+154                                                 fv = 243526                                                 fv = 1.340781e+154                                                 fv = 7.934843                                                 fv = 2793189151                                                 fv = 176634.6                                                 fv = 1470842                                                 fv = 19.89768                                                 fv = 10.9658                                                 fv = 7.35601                                                 fv = 13.17117                                                 fv = 11.13415                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.24048                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 222.8913                                                 fv = 1.923471e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.07723                                                 fv = 12.41395                                                 fv = 12.41395                                                 fv = 14.3956                                                 fv = 9.507356                                                 fv = 11.61938                                                 fv = 12.21873                                                 fv = 11.61938                                                 fv = 13.2263                                                 fv = 11.81194                                                 fv = 9.241721                                                 fv = 7.35601                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.990181                                                 fv = 10.43478                                                 fv = 1.340781e+154                                                 fv = 8.16885                                                 fv = 12.06941                                                 fv = 12.27563                                                 fv = 8.096063                                                 fv = 28.21307                                                 fv = 12.93814                                                 fv = 75154.36                                                 fv = 7.35601                                                 fv = 7.35601                                                 fv = 9.914049                                                 fv = 11.61938                                                 fv = 24084.29                                                 fv = 20.26018                                                 fv = 7.35601                                                 fv = 7.719525                                                 fv = 8.16885                                                 fv = 7.35601                                                 fv = 10.16289                                                 fv = 1.340781e+154                                                 fv = 9.990181                                                 fv = 17.65294                                                 fv = 14.21018                                                 fv = 8.641537                                                 fv = 9.486884                                                 fv = 13.08881                                                 fv = 7.35601                                                 fv = 7.965754                                                 fv = 9.849785                                                 fv = 11.81596                                                 fv = 9.468597                                                 fv = 8.492163                                                 fv = 9.701281                                                 fv = 9.49069                                                 fv = 7.35601                                                 fv = 7.35601                                                 fv = 7.35601                                                 fv = 9.373735                                                 fv = 7.35601                                                 fv = 13.2263                                                 fv = 7.35601 run # 2: f=7.356010433
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 690
#> best    : 3.743500413
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 5 1 1 1 0 0 0 2 4 10 1 1 2 1 10 1 1 ) f(x) = 3.743500413
#> 
#>                                                 fv = 3.7435                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.729779