Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 12.0919                                                 fv = 50.26729                                                 fv = 13.69581                                                 fv = 59.58967                                                 fv = 42479.25                                                 fv = 12.3996                                                 fv = 73943874                                                 fv = 1852342                                                 fv = 11.00467                                                 fv = 40.19309                                                 fv = 1.340781e+154                                                 fv = 29.14554                                                 fv = 36.42412                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.73974                                                 fv = 15447.52                                                 fv = 11.75964                                                 fv = 9.648038                                                 fv = 1.340781e+154                                                 fv = 19.43186                                                 fv = 17.23286                                                 fv = 79.92842                                                 fv = 1.340781e+154                                                 fv = 11.84829                                                 fv = 58.27137                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.3469                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 149.8931                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 974.2821                                                 fv = 27.86939                                                 fv = 15.98333                                                 fv = 91.08043                                                 fv = 9.700303                                                 fv = 121.7204                                                 fv = 159.0957                                                 fv = 17.59409                                                 fv = 9.648038                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.49902                                                 fv = 1.340781e+154                                                 fv = 22.37161                                                 fv = 14.48965 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.41855                                                 fv = 35.33333                                                 fv = 10.57918                                                 fv = 1.340781e+154                                                 fv = 90.404                                                 fv = 16.59121                                                 fv = 19.75795                                                 fv = 16.75807                                                 fv = 78.9073                                                 fv = 12.84432                                                 fv = 9.091928                                                 fv = 1.340781e+154                                                 fv = 13.10983                                                 fv = 5.705597                                                 fv = 1.340781e+154                                                 fv = 15.69188                                                 fv = 8.530632                                                 fv = 9.796513                                                 fv = 5.029534                                                 fv = 1.340781e+154                                                 fv = 16.56635                                                 fv = 5.029534                                                 fv = 4.685685                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.959825                                                 fv = 4.802472                                                 fv = 4.392843                                                 fv = 1.340781e+154                                                 fv = 4.392843                                                 fv = 1.340781e+154                                                 fv = 7.209862                                                 fv = 4.112657                                                 fv = 1.340781e+154                                                 fv = 7.40128                                                 fv = 5.532265                                                 fv = 4.112657                                                 fv = 1.340781e+154                                                 fv = 4.14052                                                 fv = 7.986825                                                 fv = 5.123899                                                 fv = 4.112657                                                 fv = 4.112657                                                 fv = 4.332653                                                 fv = 5.376786                                                 fv = 4.866846                                                 fv = 5.085611                                                 fv = 6.433079                                                 fv = 6.239406                                                 fv = 5.373378                                                 fv = 4.112657                                                 fv = 6.755417                                                 fv = 1.340781e+154                                                 fv = 7.708609                                                 fv = 7.736102                                                 fv = 4.112657                                                 fv = 4.112657                                                 fv = 6.167676                                                 fv = 5.886097                                                 fv = 63.66339                                                 fv = 17.05215                                                 fv = 13.57497                                                 fv = 4.128026                                                 fv = 1.340781e+154                                                 fv = 4.364849                                                 fv = 4.491788                                                 fv = 4.783923                                                 fv = 4.877585                                                 fv = 4.402792                                                 fv = 6.180043                                                 fv = 7.586965                                                 fv = 4.96233                                                 fv = 15.11561                                                 fv = 1.340781e+154                                                 fv = 4.862003                                                 fv = 5.085611                                                 fv = 4.995829                                                 fv = 9.919904                                                 fv = 5.18244                                                 fv = 5.108664 run # 0: f=4.112656896
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.377731                                                 fv = 32.95755                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33.16016                                                 fv = 7.504362                                                 fv = 10.0053                                                 fv = 5.672277                                                 fv = 8.706334                                                 fv = 4.421528                                                 fv = 4.421528                                                 fv = 8.435232                                                 fv = 4.299523                                                 fv = 9.183748                                                 fv = 1.340781e+154                                                 fv = 4.299523                                                 fv = 4.337761                                                 fv = 4.337761                                                 fv = 6.697736                                                 fv = 4.461599                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.337761                                                 fv = 1.340781e+154                                                 fv = 5.184721                                                 fv = 1.340781e+154                                                 fv = 4.495778                                                 fv = 4.337761                                                 fv = 1.340781e+154                                                 fv = 8.526316e+13                                                 fv = 1385774                                                 fv = 27332.89                                                 fv = 4.299523                                                 fv = 4.299523                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.896469                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.299523                                                 fv = 4.299523                                                 fv = 4.299523                                                 fv = 4.299523                                                 fv = 4.075305                                                 fv = 58.25223                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.687362                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.16645                                                 fv = 33.75131                                                 fv = 132.0655                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.04284                                                 fv = 4.075305                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 55.52076                                                 fv = 1.340781e+154                                                 fv = 4.075305                                                 fv = 6.899106                                                 fv = 4.075305                                                 fv = 4.16645                                                 fv = 1.340781e+154                                                 fv = 4.983811                                                 fv = 8.976743                                                 fv = 1.340781e+154                                                 fv = 4.050121                                                 fv = 4.050121                                                 fv = 1.340781e+154                                                 fv = 27.04284                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.050121                                                 fv = 9.356108                                                 fv = 4.461599                                                 fv = 9.356108                                                 fv = 5.841752                                                 fv = 9.356108                                                 fv = 9.356108                                                 fv = 11.78386                                                 fv = 1.340781e+154                                                 fv = 8.630279                                                 fv = 72.56697                                                 fv = 11.50504                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.04284                                                 fv = 1.340781e+154                                                 fv = 3656.003                                                 fv = 9411.276                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.33014                                                 fv = 4.703749                                                 fv = 10.9292                                                 fv = 21.0063                                                 fv = 4.272725                                                 fv = 1.340781e+154                                                 fv = 7.936136                                                 fv = 1.340781e+154                                                 fv = 4.050121                                                 fv = 3.959726                                                 fv = 54.5077                                                 fv = 1.340781e+154                                                 fv = 14.2579                                                 fv = 16.10308                                                 fv = 4.760903                                                 fv = 3.38725                                                 fv = 7.534615                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.38725                                                 fv = 3.38725                                                 fv = 1.340781e+154                                                 fv = 3.38725                                                 fv = 1.340781e+154                                                 fv = 3.959726                                                 fv = 1.340781e+154                                                 fv = 4.050121                                                 fv = 4.050121                                                 fv = 4.272725                                                 fv = 5.47142                                                 fv = 11.46613                                                 fv = 4.211928                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.84945                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.2579                                                 fv = 1.340781e+154                                                 fv = 4.096754                                                 fv = 1.340781e+154                                                 fv = 4.026235                                                 fv = 1.340781e+154                                                 fv = 6.26776                                                 fv = 4.236032                                                 fv = 3.818324                                                 fv = 1.340781e+154                                                 fv = 3.38725                                                 fv = 6.574767                                                 fv = 3.38725                                                 fv = 3.38725                                                 fv = 7.534615                                                 fv = 4.045244                                                 fv = 3.38725                                                 fv = 5.11238                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.2579                                                 fv = 3.38725                                                 fv = 5.841752                                                 fv = 5.841752                                                 fv = 4.299523                                                 fv = 1.340781e+154                                                 fv = 4.045244                                                 fv = 3.38725                                                 fv = 4.096754                                                 fv = 5.47142                                                 fv = 3.38725                                                 fv = 4.249207 run # 1: f=3.387250046
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 43.35922                                                 fv = 32.95755                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.16625                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.83179                                                 fv = 1.340781e+154                                                 fv = 12.73584                                                 fv = 11.07048                                                 fv = 1.340781e+154                                                 fv = 48710149612                                                 fv = 10.67464                                                 fv = 7.832544                                                 fv = 93.98426                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.74137                                                 fv = 1.340781e+154                                                 fv = 7.832544                                                 fv = 19.65546                                                 fv = 1.340781e+154                                                 fv = 7.952321                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 672.7298                                                 fv = 9.738162                                                 fv = 1.340781e+154                                                 fv = 8.659117                                                 fv = 580518.5                                                 fv = 1.340781e+154                                                 fv = 6.009362                                                 fv = 40449451                                                 fv = 1.340781e+154                                                 fv = 13.63906                                                 fv = 1.340781e+154                                                 fv = 158.6187                                                 fv = 1.340781e+154                                                 fv = 288149.4                                                 fv = 22.34269                                                 fv = 36.50093                                                 fv = 58.57986                                                 fv = 158.6187                                                 fv = 1.340781e+154                                                 fv = 158.6187                                                 fv = 1.340781e+154                                                 fv = 1.450861                                                 fv = 1.340781e+154                                                 fv = 1395072                                                 fv = 1.340781e+154                                                 fv = 284841281363                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 284841281363                                                 fv = 1.340781e+154                                                 fv = 1.450861                                                 fv = 6.860674                                                 fv = 40.98563                                                 fv = 100215822                                                 fv = 100215822                                                 fv = 92296987                                                 fv = 326167810222                                                 fv = 68.12544                                                 fv = 3.765558e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 498347478739                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.66012                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.14132                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.21141                                                 fv = 1.340781e+154                                                 fv = 312.2075                                                 fv = 312.2075                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 237385439857                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.119236e+12                                                 fv = 1.340781e+154                                                 fv = 1.166177e+13                                                 fv = 1.450861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.83179                                                 fv = 1.340781e+154                                                 fv = 11.63743                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4312.733                                                 fv = 1.450861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.450861                                                 fv = 1.450861                                                 fv = 1.450861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3857196                                                 fv = 11.84016                                                 fv = 10.83179                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.40745                                                 fv = 1.450861                                                 fv = 1.450861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.450861                                                 fv = 1.450861                                                 fv = 1.450861                                                 fv = 6.639864 run # 2: f=1.45086127
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 600
#> best    : 1.45086127
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 2 4 2 0 0 0 0 0 1 1 1 1 10 3 1 7 1 1 ) f(x) = 1.45086127
#> 
#>                                                 fv = 1.450861                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (3): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Warning:  optimal segment equals search maximum (10): rerun with larger segments.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 237.5691