Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 20.09146                                                 fv = 22.15501                                                 fv = 16.93222                                                 fv = 2.399634e+14                                                 fv = 1.340781e+154                                                 fv = 17.48407                                                 fv = 16.96324                                                 fv = 15710084                                                 fv = 1.340781e+154                                                 fv = 36.163                                                 fv = 518.5488                                                 fv = 1.340781e+154                                                 fv = 909.2337                                                 fv = 1.340781e+154                                                 fv = 138.5503                                                 fv = 1.340781e+154                                                 fv = 16.96324                                                 fv = 16.96324                                                 fv = 1.340781e+154                                                 fv = 12212.33                                                 fv = 1.340781e+154                                                 fv = 101.8221                                                 fv = 13392107                                                 fv = 1964.232                                                 fv = 1.284821e+13                                                 fv = 16.93222                                                 fv = 55930.83                                                 fv = 4136756                                                 fv = 21.15349                                                 fv = 36.25058                                                 fv = 16.93222                                                 fv = 1.340781e+154                                                 fv = 16.70744                                                 fv = 12369.05                                                 fv = 1.340781e+154                                                 fv = 13.99388                                                 fv = 11.8987                                                 fv = 1.340781e+154                                                 fv = 11.8987                                                 fv = 8.490519                                                 fv = 1.340781e+154                                                 fv = 15.73559                                                 fv = 19898087692                                                 fv = 1.340781e+154                                                 fv = 11.31493                                                 fv = 8.278558                                                 fv = 1.340781e+154                                                 fv = 7.41545                                                 fv = 35857.51                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 631.8321                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3119119621                                                 fv = 7.41545                                                 fv = 6.974194                                                 fv = 12.95997                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.55013                                                 fv = 1.340781e+154                                                 fv = 21.32039                                                 fv = 9.494684                                                 fv = 22.13334                                                 fv = 1.340781e+154                                                 fv = 18.44993                                                 fv = 11.06262                                                 fv = 11.06262                                                 fv = 5.659623                                                 fv = 1039.867                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.39237                                                 fv = 6.903278                                                 fv = 6.903278                                                 fv = 5.659623                                                 fv = 4.974922                                                 fv = 8.943529                                                 fv = 1.340781e+154                                                 fv = 7.976591                                                 fv = 4.974922                                                 fv = 10.25971                                                 fv = 49.70108                                                 fv = 18.72904                                                 fv = 5.278079                                                 fv = 10.43791                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.873713                                                 fv = 9.451369                                                 fv = 6.631351                                                 fv = 6.076129                                                 fv = 9.451369                                                 fv = 1.340781e+154                                                 fv = 18.82232                                                 fv = 5.278079                                                 fv = 18.58316                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 217.3322                                                 fv = 29.29618                                                 fv = 9.965756                                                 fv = 60.99716                                                 fv = 62721.87                                                 fv = 5.893778                                                 fv = 11.53289                                                 fv = 10.25971                                                 fv = 7.057625                                                 fv = 4.974922                                                 fv = 6.29085                                                 fv = 4.82273                                                 fv = 98.07733                                                 fv = 1.340781e+154                                                 fv = 40.18601                                                 fv = 17241.77                                                 fv = 19.29005                                                 fv = 13.06598                                                 fv = 10.66015                                                 fv = 4.82273                                                 fv = 4.82273                                                 fv = 11.4691                                                 fv = 5.807146                                                 fv = 5.807146                                                 fv = 21.46708                                                 fv = 6.29085                                                 fv = 12.71795                                                 fv = 9.116671                                                 fv = 24.61708                                                 fv = 4.82273                                                 fv = 23.19896                                                 fv = 8.105311                                                 fv = 3342608012                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2201.473                                                 fv = 80.48867                                                 fv = 19510.84                                                 fv = 4.82273                                                 fv = 5.476044                                                 fv = 5.145398                                                 fv = 7.183381                                                 fv = 7.063579                                                 fv = 7.184707                                                 fv = 10.00518                                                 fv = 5.893778                                                 fv = 1.340781e+154                                                 fv = 8.693407                                                 fv = 4.82273                                                 fv = 4.82273                                                 fv = 4.82273                                                 fv = 7.141664                                                 fv = 4.82273                                                 fv = 5.468108                                                 fv = 4.82273                                                 fv = 13.40501                                                 fv = 5895.466                                                 fv = 13.72783                                                 fv = 1.340781e+154                                                 fv = 5.180712                                                 fv = 10.25971                                                 fv = 12.22745                                                 fv = 5.714075                                                 fv = 6.65242                                                 fv = 6.714781                                                 fv = 5.180712                                                 fv = 4.82273                                                 fv = 6.275854                                                 fv = 1.340781e+154                                                 fv = 7.698945                                                 fv = 6.623232                                                 fv = 11.59952 run # 0: f=4.822729649
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.916061                                                 fv = 31.53425                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.835553                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.252341                                                 fv = 7.252341                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.252341                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.505042                                                 fv = 69811.73                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 191.5584                                                 fv = 7.359105                                                 fv = 64740741                                                 fv = 7.252341                                                 fv = 7.505042                                                 fv = 7.252341                                                 fv = 7.252341                                                 fv = 1.340781e+154                                                 fv = 98.21679                                                 fv = 7.252341                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 610232.4                                                 fv = 7.696083                                                 fv = 7.252341                                                 fv = 11.32267                                                 fv = 1.340781e+154                                                 fv = 169.3695                                                 fv = 33.52685                                                 fv = 7.252341                                                 fv = 7.252341                                                 fv = 7.252341                                                 fv = 7.252341                                                 fv = 1.340781e+154                                                 fv = 63.57066                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16717.03                                                 fv = 9.833987                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.569375                                                 fv = 17.80079                                                 fv = 9.833987                                                 fv = 11.89784                                                 fv = 1.340781e+154                                                 fv = 17.61817                                                 fv = 1.340781e+154                                                 fv = 17.07894                                                 fv = 7.252341                                                 fv = 8.824201                                                 fv = 1.340781e+154                                                 fv = 6.5396                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.4939                                                 fv = 8.437376                                                 fv = 6.5396                                                 fv = 1.340781e+154                                                 fv = 8.707406                                                 fv = 7.164906                                                 fv = 6.5396                                                 fv = 1.340781e+154                                                 fv = 8.366803                                                 fv = 6.5396                                                 fv = 6.5396                                                 fv = 7.296068                                                 fv = 1.340781e+154                                                 fv = 6.940983                                                 fv = 6.940983                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.684122                                                 fv = 44262.3                                                 fv = 1.340781e+154                                                 fv = 16.4939                                                 fv = 16.4939                                                 fv = 5.684122                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.634375                                                 fv = 7.634375                                                 fv = 6.591914                                                 fv = 1.340781e+154                                                 fv = 10.70692                                                 fv = 19.38596                                                 fv = 1.340781e+154                                                 fv = 17.84525                                                 fv = 5.513636                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.513636                                                 fv = 5.513636                                                 fv = 101.2303                                                 fv = 55.24644                                                 fv = 5.513636                                                 fv = 17.69821                                                 fv = 24.9022                                                 fv = 7.328139                                                 fv = 29.54742                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.926232                                                 fv = 8.926232                                                 fv = 8.926232                                                 fv = 8.712745                                                 fv = 5.513636                                                 fv = 68.66928                                                 fv = 7.714003                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.824201                                                 fv = 118.1454                                                 fv = 225265.5                                                 fv = 15.63651                                                 fv = 52.15842                                                 fv = 15.63651                                                 fv = 15.63651                                                 fv = 3.271943                                                 fv = 558876.6                                                 fv = 1.340781e+154                                                 fv = 11.64067                                                 fv = 1.340781e+154                                                 fv = 9.911012                                                 fv = 7.505042                                                 fv = 14.07292                                                 fv = 8.503957                                                 fv = 7.898957                                                 fv = 3.271943                                                 fv = 3.271943                                                 fv = 8.783323                                                 fv = 3.271943                                                 fv = 3.271943                                                 fv = 3.271943                                                 fv = 11.5854                                                 fv = 34.62274                                                 fv = 15.82947                                                 fv = 31.06015                                                 fv = 1.340781e+154                                                 fv = 3.271943                                                 fv = 1.340781e+154                                                 fv = 3.271943                                                 fv = 8.250084                                                 fv = 3.436033                                                 fv = 297.2646                                                 fv = 1.340781e+154                                                 fv = 869.2293                                                 fv = 84628.84                                                 fv = 537.7655                                                 fv = 1.340781e+154                                                 fv = 301.2647                                                 fv = 59385.83                                                 fv = 1.340781e+154                                                 fv = 20.72345                                                 fv = 9.52147                                                 fv = 1.340781e+154                                                 fv = 12.79985                                                 fv = 81.33312                                                 fv = 4.359851                                                 fv = 17.49714                                                 fv = 14.915                                                 fv = 36.33407                                                 fv = 6.071579 run # 1: f=3.271943
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 48.06119                                                 fv = 12.34195                                                 fv = 31.53425                                                 fv = 30471.51                                                 fv = 1.340781e+154                                                 fv = 138056.8                                                 fv = 12.34195                                                 fv = 7.283694                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.283694                                                 fv = 1.340781e+154                                                 fv = 9.275405                                                 fv = 14775367                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 362.053                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1358.334                                                 fv = 8.747355                                                 fv = 8.390613                                                 fv = 51299606                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.586528                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.093145                                                 fv = 8.49152                                                 fv = 1.340781e+154                                                 fv = 7.374698                                                 fv = 1.340781e+154                                                 fv = 7.283694                                                 fv = 1.340781e+154                                                 fv = 7.283694                                                 fv = 7.283694                                                 fv = 8.812021                                                 fv = 11926917 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 28.63851                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1077.531                                                 fv = 1.340781e+154                                                 fv = 51299606                                                 fv = 27.05446                                                 fv = 7.283694                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.965198                                                 fv = 1.340781e+154                                                 fv = 7.916477                                                 fv = 1.340781e+154                                                 fv = 24.5607                                                 fv = 7.283694                                                 fv = 7.749121                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5666762                                                 fv = 156.8503                                                 fv = 7.283694                                                 fv = 7.283694                                                 fv = 11.28834                                                 fv = 1.340781e+154                                                 fv = 6.392869                                                 fv = 37959.86                                                 fv = 1.340781e+154                                                 fv = 13.04132                                                 fv = 8665.412                                                 fv = 1.340781e+154                                                 fv = 44.47541                                                 fv = 75.79723                                                 fv = 1.340781e+154                                                 fv = 6.392869                                                 fv = 6.392869                                                 fv = 9.658147                                                 fv = 9.264952                                                 fv = 13.22129                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.08303                                                 fv = 1.340781e+154                                                 fv = 20.08303                                                 fv = 7.064307                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.899984                                                 fv = 23.60543                                                 fv = 14.14941                                                 fv = 19.84965                                                 fv = 1.340781e+154                                                 fv = 6.913727                                                 fv = 19.44669                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.392869                                                 fv = 1.340781e+154                                                 fv = 6.392869                                                 fv = 6.392869                                                 fv = 52.78                                                 fv = 69.42111                                                 fv = 84.06339                                                 fv = 115.2663                                                 fv = 1.340781e+154                                                 fv = 7.929768                                                 fv = 6.392869                                                 fv = 8.965198                                                 fv = 2488.331                                                 fv = 9.137266                                                 fv = 90.45478                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.45723                                                 fv = 6.395313                                                 fv = 7.334079                                                 fv = 15.71893                                                 fv = 6.392869                                                 fv = 6.941861                                                 fv = 21.97664                                                 fv = 6.392869                                                 fv = 1.340781e+154                                                 fv = 6.392869                                                 fv = 6.392869                                                 fv = 6.70568 run # 2: f=6.392868631
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 612
#> best    : 3.271943
#> worst   : 1.340780793e+154
#> solution: x = ( 1 0 0 5 1 2 2 0 0 0 1 2 3 1 1 1 3 1 1 1 ) f(x) = 3.271943
#> 
#>                                                 fv = 3.271943                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 226.3169