Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 38.76786                                                 fv = 15.89503                                                 fv = 1.340781e+154                                                 fv = 23.65688                                                 fv = 1.340781e+154                                                 fv = 134.2174                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 244.6308                                                 fv = 15.89503                                                 fv = 1.340781e+154                                                 fv = 9235.046                                                 fv = 134.2174                                                 fv = 1.340781e+154                                                 fv = 65.99104                                                 fv = 65.99104                                                 fv = 4390.027                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.49009                                                 fv = 34.96644                                                 fv = 1.340781e+154                                                 fv = 12.18462                                                 fv = 408.5987                                                 fv = 4.980988e+14                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 76.53577                                                 fv = 23.6094                                                 fv = 1.340781e+154                                                 fv = 2203.142                                                 fv = 2967.762                                                 fv = 335.2703                                                 fv = 14.15987                                                 fv = 19.37586                                                 fv = 13.95985                                                 fv = 130.2033                                                 fv = 37.70415                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6602.642                                                 fv = 1.340781e+154                                                 fv = 72.08731                                                 fv = 6.304901                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24249050                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24249050                                                 fv = 11565712                                                 fv = 103301.4                                                 fv = 6.199172                                                 fv = 11.99672                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 67.14189                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.199172                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.41772                                                 fv = 67.14189                                                 fv = 9.813455                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.4075                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.828407                                                 fv = 66881.16                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 346.2537                                                 fv = 1.340781e+154                                                 fv = 108317.9                                                 fv = 43.16997                                                 fv = 298549.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38.40207                                                 fv = 4.614624                                                 fv = 3.985686                                                 fv = 3.985686                                                 fv = 4.673647                                                 fv = 4.673647                                                 fv = 1.340781e+154                                                 fv = 3.828407                                                 fv = 8.105375                                                 fv = 1.340781e+154                                                 fv = 683.7873                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 780.3855                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 72.83109                                                 fv = 990.2266                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.828407                                                 fv = 990.2266                                                 fv = 78238.68                                                 fv = 1.340781e+154                                                 fv = 3.828407                                                 fv = 398640424492                                                 fv = 3.985686                                                 fv = 3.828407                                                 fv = 250.1714                                                 fv = 4.584769                                                 fv = 3.985686                                                 fv = 217.4343                                                 fv = 3.828407                                                 fv = 3.828407                                                 fv = 1.340781e+154                                                 fv = 7.161349                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 59.19854                                                 fv = 1.340781e+154                                                 fv = 36.70216                                                 fv = 4.584769                                                 fv = 3910.585                                                 fv = 28.85349                                                 fv = 26.52587                                                 fv = 42.56266                                                 fv = 28.85349                                                 fv = 4.619483                                                 fv = 3.828407                                                 fv = 3.828407                                                 fv = 1.340781e+154                                                 fv = 3.985686                                                 fv = 4.08648                                                 fv = 3.828407                                                 fv = 475399.4                                                 fv = 335.0669                                                 fv = 5.362417                                                 fv = 3.828407                                                 fv = 6.156836                                                 fv = 1.340781e+154                                                 fv = 70.04066                                                 fv = 1.340781e+154                                                 fv = 3.828407                                                 fv = 4.841627                                                 fv = 7.823844                                                 fv = 1.340781e+154                                                 fv = 4.650784                                                 fv = 5.448784                                                 fv = 92.3825                                                 fv = 28.85349                                                 fv = 297.0867                                                 fv = 6.428737                                                 fv = 152528175                                                 fv = 6.319503 run # 0: f=3.828406533
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.641217                                                 fv = 31.71315                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.05782                                                 fv = 7.871414                                                 fv = 7.939288                                                 fv = 11.53566                                                 fv = 1.340781e+154                                                 fv = 7.641217                                                 fv = 8.268282                                                 fv = 23766.71                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.939288                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.123571                                                 fv = 1.340781e+154                                                 fv = 269.4617                                                 fv = 6.10734                                                 fv = 28.75697                                                 fv = 12255.56                                                 fv = 1.340781e+154                                                 fv = 184183.3                                                 fv = 151.8152                                                 fv = 26590.63                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.1568                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.59384                                                 fv = 61035.75                                                 fv = 46.81548                                                 fv = 1524176323                                                 fv = 6.10734                                                 fv = 6.10734                                                 fv = 1.340781e+154                                                 fv = 140448.4                                                 fv = 79.15957                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 52.39636                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.97031                                                 fv = 12.47274                                                 fv = 1.340781e+154                                                 fv = 6.03133                                                 fv = 9.610951                                                 fv = 1.340781e+154                                                 fv = 29.32316                                                 fv = 6.03133                                                 fv = 27.34309                                                 fv = 230.4888                                                 fv = 30.17296                                                 fv = 5.825389                                                 fv = 965496.9                                                 fv = 1.340781e+154                                                 fv = 25.09936                                                 fv = 5.825389                                                 fv = 10.43097                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.825389                                                 fv = 1.340781e+154                                                 fv = 29.07598                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.465625                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 383.7073                                                 fv = 5.341083                                                 fv = 10.90161                                                 fv = 1.340781e+154                                                 fv = 25.09936                                                 fv = 1.340781e+154                                                 fv = 8.028654                                                 fv = 14.80643                                                 fv = 14.80643                                                 fv = 1.340781e+154                                                 fv = 13.53056                                                 fv = 23.67236                                                 fv = 48.95456                                                 fv = 1.340781e+154                                                 fv = 13.53056                                                 fv = 1.340781e+154                                                 fv = 7.656534                                                 fv = 198.4921                                                 fv = 96430.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.32316                                                 fv = 9.262039                                                 fv = 1.340781e+154                                                 fv = 135703.6                                                 fv = 34.80831                                                 fv = 6.266545                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 2070.765                                                 fv = 1.340781e+154                                                 fv = 25.09936                                                 fv = 482292.2                                                 fv = 1.340781e+154                                                 fv = 634.8419                                                 fv = 412802690                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 15.24223                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 61894.28                                                 fv = 13.77524                                                 fv = 13.77524                                                 fv = 1.340781e+154                                                 fv = 58.69152                                                 fv = 11.06004                                                 fv = 3.070088                                                 fv = 5.583526                                                 fv = 3.070088                                                 fv = 1.340781e+154                                                 fv = 36.10393                                                 fv = 1.340781e+154                                                 fv = 34.90533                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 2.561544                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 407.7278                                                 fv = 6.446347                                                 fv = 2.561544                                                 fv = 35.39074                                                 fv = 23.80023                                                 fv = 14.43269                                                 fv = 18.03037                                                 fv = 11.56643                                                 fv = 1.340781e+154                                                 fv = 42.85388                                                 fv = 7.313647                                                 fv = 11.44013                                                 fv = 13390.55                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.49312                                                 fv = 82.1727                                                 fv = 10.61255                                                 fv = 3.19497                                                 fv = 11.6012                                                 fv = 1.340781e+154                                                 fv = 20.25113                                                 fv = 2.561544                                                 fv = 2.561544                                                 fv = 3.205725                                                 fv = 7.783619                                                 fv = 2.561544                                                 fv = 1.340781e+154                                                 fv = 4.0541                                                 fv = 1.340781e+154                                                 fv = 2.561544                                                 fv = 2.561544                                                 fv = 2.561544                                                 fv = 534723.7                                                 fv = 2.561544                                                 fv = 103.6927                                                 fv = 167.1561                                                 fv = 134.7553                                                 fv = 3.88571                                                 fv = 12.30168                                                 fv = 2.561544                                                 fv = 1.340781e+154                                                 fv = 12.20027                                                 fv = 87.85561                                                 fv = 3.048683                                                 fv = 2.561544                                                 fv = 29.06613                                                 fv = 5.198388                                                 fv = 4.715746                                                 fv = 5.583526                                                 fv = 5.583526                                                 fv = 9.464545                                                 fv = 18.03037 run # 1: f=2.561544086
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19376.5                                                 fv = 31.71315                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.49763                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.13888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.80272                                                 fv = 4154.597                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.79467                                                 fv = 11.59367                                                 fv = 3422759                                                 fv = 10.93987                                                 fv = 1.340781e+154                                                 fv = 359.0914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.35483                                                 fv = 17.12944                                                 fv = 5.796401e+13                                                 fv = 16.94532                                                 fv = 10.79467                                                 fv = 10.79467                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 543.5079                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 130100.8                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.79467                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33.16321                                                 fv = 331.7853                                                 fv = 21.29189                                                 fv = 13.10062                                                 fv = 13.4367                                                 fv = 1.340781e+154                                                 fv = 51.39668                                                 fv = 15.29222                                                 fv = 1.340781e+154                                                 fv = 25.49763                                                 fv = 1.340781e+154                                                 fv = 10.79467                                                 fv = 10.79467                                                 fv = 1.340781e+154                                                 fv = 10.79467                                                 fv = 10.79467                                                 fv = 10.79467                                                 fv = 9.985674                                                 fv = 10804948712                                                 fv = 1.340781e+154                                                 fv = 11.80272                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.985674                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.985674                                                 fv = 9.985674                                                 fv = 9.947446                                                 fv = 11.80272                                                 fv = 12.13888                                                 fv = 11.15991                                                 fv = 1.340781e+154                                                 fv = 10.72962                                                 fv = 10.72962                                                 fv = 17.06201                                                 fv = 11.77821                                                 fv = 10.72962                                                 fv = 9.947446                                                 fv = 9.947446                                                 fv = 9.947446                                                 fv = 11.09987                                                 fv = 10.72962                                                 fv = 11724006009                                                 fv = 25.49763                                                 fv = 290.5102                                                 fv = 1279.023                                                 fv = 13.02256                                                 fv = 1.340781e+154                                                 fv = 37.9802                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.56686                                                 fv = 15.07839                                                 fv = 19.99157                                                 fv = 12.89444                                                 fv = 10.13644                                                 fv = 40.09201                                                 fv = 20.19898                                                 fv = 876.9933                                                 fv = 21.83685                                                 fv = 9.947446                                                 fv = 1.340781e+154                                                 fv = 9.947446                                                 fv = 9.947446                                                 fv = 1.340781e+154                                                 fv = 16.17063                                                 fv = 9.947446                                                 fv = 9.947446                                                 fv = 9328373109                                                 fv = 14.09353                                                 fv = 1.340781e+154                                                 fv = 11.80272                                                 fv = 10.41999                                                 fv = 11.2678                                                 fv = 17.17851                                                 fv = 10.85877                                                 fv = 12.8758                                                 fv = 171.959                                                 fv = 9.947446                                                 fv = 15.07839                                                 fv = 9.947446                                                 fv = 1.340781e+154                                                 fv = 12.89444                                                 fv = 9.947446                                                 fv = 13.21493                                                 fv = 31.81743 run # 3: f=9.947446087
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 653
#> best    : 2.561544086
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 7 0 0 0 1 0 0 1 1 4 1 2 10 2 3 1 2 ) f(x) = 2.561544086
#> 
#>                                                 fv = 2.561544                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.760489