Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 25.34264                                                 fv = 26.04026                                                 fv = 1.340781e+154                                                 fv = 266.1723                                                 fv = 9.922622                                                 fv = 272.2358                                                 fv = 11884.88                                                 fv = 132.0218                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 30.71732                                                 fv = 1.340781e+154                                                 fv = 132.0218                                                 fv = 132.0218                                                 fv = 1.340781e+154                                                 fv = 9.637012                                                 fv = 78.07609                                                 fv = 32.60059                                                 fv = 1.340781e+154                                                 fv = 36.52989                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.66059                                                 fv = 37.04647                                                 fv = 11.64245                                                 fv = 1.340781e+154                                                 fv = 9763289526                                                 fv = 32.60059                                                 fv = 1.340781e+154                                                 fv = 634.5175                                                 fv = 94.18715                                                 fv = 25088.51                                                 fv = 1.340781e+154                                                 fv = 3181.854                                                 fv = 9.637012                                                 fv = 1.340781e+154                                                 fv = 92.15121                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3093.558                                                 fv = 104.87                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.35422                                                 fv = 59.15259                                                 fv = 4375.238                                                 fv = 5321.417                                                 fv = 15.599                                                 fv = 76.10131                                                 fv = 1.340781e+154                                                 fv = 14.72274                                                 fv = 14.72274                                                 fv = 25.51868                                                 fv = 28.64812                                                 fv = 9.637012                                                 fv = 14.72274                                                 fv = 95.57033                                                 fv = 1.340781e+154                                                 fv = 56.12968                                                 fv = 1.340781e+154                                                 fv = 55.73017                                                 fv = 9.637012                                                 fv = 10.92926                                                 fv = 1.340781e+154                                                 fv = 12.63016                                                 fv = 11.60758                                                 fv = 10.70522                                                 fv = 79.68913                                                 fv = 14.92365                                                 fv = 19.39673                                                 fv = 9.637012                                                 fv = 13.58023                                                 fv = 1.340781e+154                                                 fv = 27.8408                                                 fv = 26.06518                                                 fv = 29.1906                                                 fv = 77.78081                                                 fv = 9.637012                                                 fv = 9.637012                                                 fv = 9.637012                                                 fv = 1.340781e+154 run # 0: f=9.637012184
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.8547                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.8547                                                 fv = 35.43766                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31843.76                                                 fv = 9.543701                                                 fv = 41.66624                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 274.9617                                                 fv = 1.340781e+154                                                 fv = 6.813604                                                 fv = 169.054                                                 fv = 36.75164                                                 fv = 6.813604                                                 fv = 6.813604                                                 fv = 6.813604                                                 fv = 10.84025                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.98071                                                 fv = 29.82739                                                 fv = 915.9861                                                 fv = 10.49633                                                 fv = 1.340781e+154                                                 fv = 2830.682                                                 fv = 11.60406                                                 fv = 6.813604                                                 fv = 9.944082                                                 fv = 1842466537                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.31989                                                 fv = 6.813604                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.15041                                                 fv = 1.340781e+154                                                 fv = 8.320028                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.922031                                                 fv = 5.677554                                                 fv = 1.340781e+154                                                 fv = 12.15041                                                 fv = 8.069028                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 183453483                                                 fv = 6.813604                                                 fv = 5.677554                                                 fv = 5.677554                                                 fv = 6.813604                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.84025                                                 fv = 13.32122                                                 fv = 13.0343                                                 fv = 5.551386                                                 fv = 5.551386                                                 fv = 1.340781e+154                                                 fv = 10.84025                                                 fv = 5.551386                                                 fv = 5.551386                                                 fv = 505185571                                                 fv = 5.475848                                                 fv = 1.340781e+154                                                 fv = 12.84759                                                 fv = 7.642499                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.677554                                                 fv = 5.677554                                                 fv = 5.475848                                                 fv = 12.79981                                                 fv = 5.551386                                                 fv = 5.551386                                                 fv = 6.527166                                                 fv = 5.475848                                                 fv = 5.475848                                                 fv = 4.555285                                                 fv = 1.340781e+154                                                 fv = 16.12234                                                 fv = 1.340781e+154                                                 fv = 4.555285                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.555285                                                 fv = 7.305173                                                 fv = 5.475848                                                 fv = 6.073998                                                 fv = 1.340781e+154                                                 fv = 7.601719                                                 fv = 6.668405                                                 fv = 5.739746                                                 fv = 4.555285                                                 fv = 4.555285                                                 fv = 734.4201                                                 fv = 5.087123                                                 fv = 5.677554                                                 fv = 4.555285                                                 fv = 6.845542                                                 fv = 7.929308                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.12234                                                 fv = 1.340781e+154                                                 fv = 5.099795                                                 fv = 12.13719                                                 fv = 1.340781e+154                                                 fv = 4.555285                                                 fv = 6.674211                                                 fv = 4.555285                                                 fv = 4.555285                                                 fv = 4.555285                                                 fv = 4.555285                                                 fv = 4.555285                                                 fv = 1.340781e+154                                                 fv = 5.886832                                                 fv = 11.66245                                                 fv = 4.555285                                                 fv = 2339.968                                                 fv = 44.89939                                                 fv = 1.340781e+154                                                 fv = 5.97772                                                 fv = 4.555285                                                 fv = 5.990677                                                 fv = 4.773114                                                 fv = 1.340781e+154                                                 fv = 6.07347                                                 fv = 3.643078                                                 fv = 1.340781e+154                                                 fv = 4.555285                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.586528                                                 fv = 4.325474                                                 fv = 4.586528                                                 fv = 3.643078                                                 fv = 33.37771                                                 fv = 3.643078                                                 fv = 3.643078                                                 fv = 3.643078                                                 fv = 5.087123                                                 fv = 5.790328                                                 fv = 4.993171                                                 fv = 4.71571                                                 fv = 4.586528                                                 fv = 3.865854                                                 fv = 1.340781e+154                                                 fv = 3.656437                                                 fv = 4.439762                                                 fv = 3263.326                                                 fv = 1.340781e+154                                                 fv = 3.643078                                                 fv = 1.340781e+154                                                 fv = 3.966092                                                 fv = 4.83074                                                 fv = 5.269087                                                 fv = 3.865854                                                 fv = 6.136704                                                 fv = 1.340781e+154                                                 fv = 4.325474                                                 fv = 3.643078                                                 fv = 5.725023                                                 fv = 4.152327                                                 fv = 3.643078                                                 fv = 3.643078                                                 fv = 45.8547 run # 1: f=3.643078252
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 64.70387                                                 fv = 12.12833                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.359348                                                 fv = 15.50199                                                 fv = 193798789                                                 fv = 12.50278                                                 fv = 1.340781e+154                                                 fv = 34.56662                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 118.7341                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7489.606                                                 fv = 1.340781e+154                                                 fv = 1001933                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 542.3483                                                 fv = 113109977                                                 fv = 11.37391                                                 fv = 9.068946                                                 fv = 1.340781e+154                                                 fv = 14.70939                                                 fv = 93632313                                                 fv = 1.340781e+154                                                 fv = 12.43829                                                 fv = 1.340781e+154                                                 fv = 9.339068                                                 fv = 21.89109                                                 fv = 1.340781e+154                                                 fv = 16.51708                                                 fv = 9.068946                                                 fv = 1256.059                                                 fv = 8.818043                                                 fv = 17.12009                                                 fv = 8.359348                                                 fv = 8.359348                                                 fv = 8.359348                                                 fv = 8.359348                                                 fv = 264.9481                                                 fv = 8.359348                                                 fv = 4502.269                                                 fv = 11.64213                                                 fv = 1.340781e+154                                                 fv = 17.07448                                                 fv = 329091                                                 fv = 9.212093                                                 fv = 8.761768                                                 fv = 10.63864                                                 fv = 13.24787                                                 fv = 8.930687                                                 fv = 9.068946                                                 fv = 5.442646                                                 fv = 1086802                                                 fv = 1.340781e+154                                                 fv = 12.92038                                                 fv = 12.92038                                                 fv = 30.40535                                                 fv = 7.415649                                                 fv = 5.442646                                                 fv = 132.0238                                                 fv = 6.847136                                                 fv = 5.442646                                                 fv = 11.39355                                                 fv = 1.340781e+154                                                 fv = 6.951632                                                 fv = 80.91461                                                 fv = 30.40535                                                 fv = 6.847136                                                 fv = 6.847136                                                 fv = 29.29442                                                 fv = 5.308366                                                 fv = 5.308366                                                 fv = 1.340781e+154                                                 fv = 8.532546                                                 fv = 5.308366                                                 fv = 5.308366                                                 fv = 11.46588                                                 fv = 40.99138                                                 fv = 5.308366                                                 fv = 6.055332                                                 fv = 9.835831                                                 fv = 1.340781e+154                                                 fv = 5.873522                                                 fv = 5.124422                                                 fv = 788.4736                                                 fv = 1.340781e+154                                                 fv = 66.93284                                                 fv = 1.340781e+154                                                 fv = 80.53936                                                 fv = 31.85029                                                 fv = 7.976432                                                 fv = 1.340781e+154                                                 fv = 5.124422                                                 fv = 5.124422                                                 fv = 6.951632                                                 fv = 9.336712                                                 fv = 5.124422                                                 fv = 5.124422                                                 fv = 1.340781e+154                                                 fv = 4.991107                                                 fv = 1.340781e+154                                                 fv = 11.46588                                                 fv = 5.68433                                                 fv = 4.991107                                                 fv = 17.15204                                                 fv = 14.54888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20868560                                                 fv = 4.991107                                                 fv = 16.37                                                 fv = 17.15204                                                 fv = 7.434902                                                 fv = 1.340781e+154                                                 fv = 13.1534                                                 fv = 8.289256                                                 fv = 15.975                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.28171                                                 fv = 1.340781e+154                                                 fv = 6300161845                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 158.655                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.268459                                                 fv = 12.06877                                                 fv = 1.340781e+154                                                 fv = 10.71285                                                 fv = 9696.414                                                 fv = 6.887371                                                 fv = 377317.9                                                 fv = 1.340781e+154                                                 fv = 4.991107                                                 fv = 46.02464                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 13.1534                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 6.417398                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 247.0171                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.143871                                                 fv = 8.809915                                                 fv = 1.340781e+154                                                 fv = 6.924617                                                 fv = 17.28462                                                 fv = 6.887371                                                 fv = 5.41609                                                 fv = 1.340781e+154                                                 fv = 4.991107                                                 fv = 44.59211                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 7.434902                                                 fv = 4.991107                                                 fv = 4.991107                                                 fv = 5.743683                                                 fv = 4.63194                                                 fv = 11.46588                                                 fv = 1.340781e+154                                                 fv = 64209.32                                                 fv = 4.900112                                                 fv = 4.63194                                                 fv = 4.804571                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 6.381395                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 8.838288                                                 fv = 5.778924                                                 fv = 5.68433                                                 fv = 1.340781e+154                                                 fv = 4.63194                                                 fv = 4.644639                                                 fv = 4.63194                                                 fv = 4.637898                                                 fv = 4.730609                                                 fv = 4.63194                                                 fv = 11.46588                                                 fv = 4.675201                                                 fv = 5.231552                                                 fv = 5.427978                                                 fv = 5.206604                                                 fv = 5.182862                                                 fv = 6.033261                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 4.943212                                                 fv = 1.340781e+154                                                 fv = 43.43295                                                 fv = 1.340781e+154                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 4.63194                                                 fv = 4.943212                                                 fv = 4.56974                                                 fv = 11.27535                                                 fv = 1.340781e+154                                                 fv = 21.4555                                                 fv = 4.56974                                                 fv = 4.251533                                                 fv = 159.3163                                                 fv = 1.340781e+154                                                 fv = 4.713806                                                 fv = 4.593064                                                 fv = 4.251533                                                 fv = 4.251533                                                 fv = 4.251533                                                 fv = 4.630059                                                 fv = 4.251533                                                 fv = 1.340781e+154                                                 fv = 8.872429                                                 fv = 4.664022                                                 fv = 4.251533                                                 fv = 4.261235                                                 fv = 4.251533                                                 fv = 4.251533                                                 fv = 4.410315                                                 fv = 7.415649                                                 fv = 4.177861                                                 fv = 8.020344                                                 fv = 1.340781e+154                                                 fv = 28.80059                                                 fv = 4.177861                                                 fv = 4.818584                                                 fv = 5.013283                                                 fv = 4.177861                                                 fv = 4.401671                                                 fv = 4.177861                                                 fv = 5.124422                                                 fv = 5.156648                                                 fv = 4.561855                                                 fv = 4.51559                                                 fv = 1.340781e+154                                                 fv = 4.315425                                                 fv = 4.177861                                                 fv = 5.811051                                                 fv = 4.557322                                                 fv = 9.418369                                                 fv = 5.174534                                                 fv = 4.177861                                                 fv = 11.46588                                                 fv = 9.784959 run # 2: f=4.177861348
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.72598                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.72598                                                 fv = 1.340781e+154                                                 fv = 16.15672                                                 fv = 18.03328                                                 fv = 1.340781e+154                                                 fv = 14.06891                                                 fv = 1.340781e+154                                                 fv = 24.60134                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1131510940                                                 fv = 14.55233                                                 fv = 16.65617                                                 fv = 1.340781e+154                                                 fv = 14.06891                                                 fv = 1.340781e+154                                                 fv = 14.06891                                                 fv = 14.06891                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2543071534                                                 fv = 16.55804                                                 fv = 14.6351                                                 fv = 45.8547                                                 fv = 37.55577                                                 fv = 1.340781e+154                                                 fv = 19.03768                                                 fv = 10.13444                                                 fv = 10912.85                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.00886                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.201831                                                 fv = 10.27912                                                 fv = 1.340781e+154                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8472.485                                                 fv = 9.526033                                                 fv = 1.340781e+154                                                 fv = 58248.71                                                 fv = 4.201831                                                 fv = 13.08275                                                 fv = 5.776596                                                 fv = 4.32379                                                 fv = 6.346202                                                 fv = 4.32379                                                 fv = 11.88248                                                 fv = 4.201831                                                 fv = 10.45705                                                 fv = 1.340781e+154                                                 fv = 22.26624                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 42.74488                                                 fv = 34.63917                                                 fv = 42.22504                                                 fv = 1.340781e+154                                                 fv = 7.702005                                                 fv = 4.201831                                                 fv = 5.646047                                                 fv = 16.17371                                                 fv = 16.17371                                                 fv = 83.60472                                                 fv = 6484415                                                 fv = 22.26624                                                 fv = 16.37214                                                 fv = 13.65465                                                 fv = 1.340781e+154                                                 fv = 4.201831                                                 fv = 11.47756                                                 fv = 21.64672                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.0116                                                 fv = 17.67984                                                 fv = 1.340781e+154                                                 fv = 9.139143                                                 fv = 680.0422                                                 fv = 6.323965                                                 fv = 5.372102                                                 fv = 1.340781e+154                                                 fv = 842.7141                                                 fv = 9.590939                                                 fv = 1.340781e+154                                                 fv = 5.00886                                                 fv = 6.053781                                                 fv = 10.56313                                                 fv = 4.201831                                                 fv = 4.201831                                                 fv = 15.2672                                                 fv = 9.482045                                                 fv = 4.201831                                                 fv = 19.89055                                                 fv = 14.95261                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 45.8547                                                 fv = 45.8547                                                 fv = 9.240365                                                 fv = 1.340781e+154                                                 fv = 7.138694                                                 fv = 4.201831                                                 fv = 5.585779                                                 fv = 23.58906                                                 fv = 4.201831                                                 fv = 5.390558                                                 fv = 15.30272                                                 fv = 8.777679                                                 fv = 4.201831                                                 fv = 4.201831                                                 fv = 5.909571                                                 fv = 4.538288                                                 fv = 4.201831                                                 fv = 6.857098                                                 fv = 10.26441                                                 fv = 10.45705                                                 fv = 12.97906                                                 fv = 40.10298 run # 3: f=4.201831194
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.329986                                                 fv = 45.8547                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38.80492                                                 fv = 22.57563                                                 fv = 12.40789                                                 fv = 20.9184                                                 fv = 17.7131                                                 fv = 40.76646                                                 fv = 1.340781e+154                                                 fv = 4.14525                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 202.2585                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 338.5557                                                 fv = 75.40205                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7164.087                                                 fv = 35.4133                                                 fv = 70151.64                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 422.9353                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.903132                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 110.1257                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.14525                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.14525                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24353.12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1459.253                                                 fv = 5.126082                                                 fv = 102348.6                                                 fv = 1674.941                                                 fv = 2516.224                                                 fv = 26880.1                                                 fv = 4.14525                                                 fv = 51121.88                                                 fv = 1.340781e+154                                                 fv = 22.25408                                                 fv = 10.53375                                                 fv = 77.3413                                                 fv = 4.14525                                                 fv = 32.58907                                                 fv = 4.14525                                                 fv = 152.3948                                                 fv = 25.54673                                                 fv = 633.4513                                                 fv = 21.61464                                                 fv = 59.57549                                                 fv = 8.412954 run # 4: f=4.145250115
#> 
#> bb eval : 819
#> best    : 3.643078252
#> worst   : 9.637012184
#> solution: x = ( 1 0 0 1 0 1 0 0 0 0 1 5 1 3 1 1 1 1 1 1 ) f(x) = 3.643078252
#> 
#>                                                 fv = 3.643078                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...          

# Score the predictions
predictions$score()
#> regr.mse 
#> 24.54423