Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 40.16584                                                 fv = 28.54445                                                 fv = 11671.6                                                 fv = 28.10977                                                 fv = 111.0642                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.58519                                                 fv = 35.84355                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.58519                                                 fv = 1.340781e+154                                                 fv = 27.26216                                                 fv = 30.13936                                                 fv = 40.57702                                                 fv = 297.2099                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 83.7705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1013.968                                                 fv = 65.4772                                                 fv = 365.181                                                 fv = 1.995073e+12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 56.88164                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.36052                                                 fv = 71.28259                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.26216                                                 fv = 1.340781e+154                                                 fv = 18.50352                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 148.4208                                                 fv = 14.52533                                                 fv = 1.340781e+154                                                 fv = 1056.283                                                 fv = 14.63032                                                 fv = 163.8367                                                 fv = 84.16141                                                 fv = 14.52533                                                 fv = 56.92841                                                 fv = 15.01165                                                 fv = 15.01165                                                 fv = 14.52533                                                 fv = 11.17663                                                 fv = 1241.453                                                 fv = 1.340781e+154                                                 fv = 16.11598                                                 fv = 32.5898                                                 fv = 18.96353                                                 fv = 13.12077                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.17663                                                 fv = 12.42723                                                 fv = 149.416                                                 fv = 48.27525                                                 fv = 44.74321                                                 fv = 69.2969                                                 fv = 31.76247                                                 fv = 69.2969                                                 fv = 69.2969                                                 fv = 1.340781e+154                                                 fv = 39.86441                                                 fv = 128.405                                                 fv = 226.9664                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.83032                                                 fv = 84324.31                                                 fv = 150.5322                                                 fv = 30.41636                                                 fv = 21.26662                                                 fv = 84.1378                                                 fv = 91.82836                                                 fv = 8494.084                                                 fv = 11.17663                                                 fv = 1.340781e+154                                                 fv = 11.17663                                                 fv = 14.27335                                                 fv = 11.17663                                                 fv = 1.340781e+154                                                 fv = 12.42723                                                 fv = 11.17663                                                 fv = 1.340781e+154                                                 fv = 17.38447                                                 fv = 42.56355                                                 fv = 12.39111                                                 fv = 1.340781e+154                                                 fv = 22.54629                                                 fv = 56.96295                                                 fv = 12.89756                                                 fv = 45.54463                                                 fv = 29.9482                                                 fv = 21.87169                                                 fv = 58.74725                                                 fv = 21.66379                                                 fv = 1.340781e+154                                                 fv = 11.17663                                                 fv = 11.17663                                                 fv = 16.17741                                                 fv = 11.76985                                                 fv = 17.26823 run # 0: f=11.17663338
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 295.5164                                                 fv = 38.35785                                                 fv = 38.35785                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38.35785                                                 fv = 284367.7                                                 fv = 38.35785                                                 fv = 12.40503                                                 fv = 25.99708                                                 fv = 1.340781e+154                                                 fv = 16.66672                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.23795                                                 fv = 40.88829                                                 fv = 23.27871                                                 fv = 9.536317                                                 fv = 13.30576                                                 fv = 13.79777                                                 fv = 1.340781e+154                                                 fv = 9.536317                                                 fv = 3269902653                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 274.0327                                                 fv = 1.340781e+154                                                 fv = 31.44965                                                 fv = 11.34913                                                 fv = 1.340781e+154                                                 fv = 4629.841                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.523244                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20344.15                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 608.4013                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 42.16864                                                 fv = 1.340781e+154                                                 fv = 786.6089                                                 fv = 188547.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.2264                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.05656                                                 fv = 11.4594                                                 fv = 11.4594                                                 fv = 14.21635                                                 fv = 8.523244                                                 fv = 1.340781e+154                                                 fv = 17.93019                                                 fv = 1.340781e+154                                                 fv = 17.3184                                                 fv = 6469498                                                 fv = 3660.136                                                 fv = 52.168                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.0072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.70739                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.38607                                                 fv = 16.01644                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.99119                                                 fv = 12.99119                                                 fv = 1.340781e+154                                                 fv = 12.54309                                                 fv = 1.340781e+154                                                 fv = 8.523244                                                 fv = 10.95582                                                 fv = 12.3098                                                 fv = 8.523244                                                 fv = 1.340781e+154                                                 fv = 19042470                                                 fv = 240989751535                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.2264                                                 fv = 172.6663                                                 fv = 1.340781e+154                                                 fv = 9.263429                                                 fv = 99449.68                                                 fv = 8.523244                                                 fv = 9.456703                                                 fv = 11.48213                                                 fv = 8.523244                                                 fv = 8.523244                                                 fv = 8.523244                                                 fv = 12.60934                                                 fv = 7.278331                                                 fv = 1.340781e+154                                                 fv = 12.52593                                                 fv = 7.278331                                                 fv = 1.340781e+154                                                 fv = 8.523244                                                 fv = 8.545583                                                 fv = 106.983                                                 fv = 9.923586                                                 fv = 6.893648                                                 fv = 1227411                                                 fv = 1.340781e+154                                                 fv = 16.99241                                                 fv = 15.37001                                                 fv = 9.096833                                                 fv = 24.06315                                                 fv = 6300.47                                                 fv = 14.54088                                                 fv = 6.893648                                                 fv = 11.9605                                                 fv = 7.165923                                                 fv = 6.893648                                                 fv = 6.893648                                                 fv = 3456922                                                 fv = 9.096833                                                 fv = 15.04735                                                 fv = 6.893648                                                 fv = 12.27616                                                 fv = 9.932407                                                 fv = 25.27675                                                 fv = 3865.123                                                 fv = 12.55979                                                 fv = 13.43023                                                 fv = 6.893648                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 1.340781e+154                                                 fv = 9.505775                                                 fv = 1.340781e+154                                                 fv = 61.9599                                                 fv = 27.36052                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 12.45925                                                 fv = 11.23849                                                 fv = 10.74178                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 10.34877                                                 fv = 12.70721                                                 fv = 12.33133                                                 fv = 69695.61                                                 fv = 11.05616                                                 fv = 11.90892                                                 fv = 6.703608                                                 fv = 1.340781e+154                                                 fv = 3091141252                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 15.41988                                                 fv = 13.32855                                                 fv = 11.76123                                                 fv = 17.58781                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 10.11609                                                 fv = 11.28399                                                 fv = 9.873334                                                 fv = 12.40861                                                 fv = 16.99241                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 24.43253                                                 fv = 18.30775                                                 fv = 1.340781e+154                                                 fv = 11.47578                                                 fv = 10.46788                                                 fv = 12.79122                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 1.340781e+154                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 11.28399                                                 fv = 11.7503                                                 fv = 9.180132                                                 fv = 9.347771                                                 fv = 10.48336                                                 fv = 10.07985                                                 fv = 6.703608                                                 fv = 15.24696                                                 fv = 15.21959                                                 fv = 1.340781e+154                                                 fv = 16.07544                                                 fv = 10.46788                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 11.59444                                                 fv = 6.703608                                                 fv = 6.703608                                                 fv = 8.491535                                                 fv = 12.77083 run # 2: f=6.703608472
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25.99708                                                 fv = 38.35785                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.36466                                                 fv = 7602.64                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.68677                                                 fv = 455987.4                                                 fv = 13.50463                                                 fv = 1.340781e+154                                                 fv = 21.13125                                                 fv = 25.72778                                                 fv = 25.78037                                                 fv = 1.340781e+154                                                 fv = 10.68677                                                 fv = 10.36466                                                 fv = 73461994977                                                 fv = 13.50463                                                 fv = 959.4594                                                 fv = 48728.06                                                 fv = 349238876                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7064334                                                 fv = 1.340781e+154                                                 fv = 10.36466                                                 fv = 21.48414                                                 fv = 10.19161                                                 fv = 1.340781e+154                                                 fv = 15.56678                                                 fv = 12.18391                                                 fv = 1.340781e+154                                                 fv = 28.65632                                                 fv = 20.71018                                                 fv = 58382.93                                                 fv = 15.49287                                                 fv = 20.91643                                                 fv = 1.340781e+154                                                 fv = 15.49287                                                 fv = 1.340781e+154                                                 fv = 203.5954                                                 fv = 22.57897                                                 fv = 12.18391                                                 fv = 1.340781e+154                                                 fv = 25.95183                                                 fv = 91322.41                                                 fv = 18.35114                                                 fv = 25.30223                                                 fv = 27.56402                                                 fv = 18.35114                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.23691                                                 fv = 1.340781e+154                                                 fv = 13.777                                                 fv = 1.340781e+154                                                 fv = 15.01323                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.349                                                 fv = 11.00798                                                 fv = 12.7821                                                 fv = 14.40534                                                 fv = 8.669426                                                 fv = 3395893                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.669426                                                 fv = 22.80428                                                 fv = 25.12843                                                 fv = 8.669426                                                 fv = 8.263252                                                 fv = 23763.23                                                 fv = 15.23691                                                 fv = 17.36548                                                 fv = 11.91975                                                 fv = 12.14208                                                 fv = 1.340781e+154                                                 fv = 11.91975                                                 fv = 19.79155                                                 fv = 8.263252                                                 fv = 13.82889                                                 fv = 8.263252                                                 fv = 21.29452                                                 fv = 8.669426                                                 fv = 9.369829                                                 fv = 6.425705                                                 fv = 1.340781e+154                                                 fv = 14.59706                                                 fv = 8.895747                                                 fv = 7.062758                                                 fv = 15.24894                                                 fv = 6.783881                                                 fv = 1.340781e+154                                                 fv = 9.658547                                                 fv = 6.386378                                                 fv = 1.340781e+154                                                 fv = 14.41056                                                 fv = 1.340781e+154                                                 fv = 562.5495                                                 fv = 143.2342                                                 fv = 6.386378                                                 fv = 9.806994                                                 fv = 1.340781e+154                                                 fv = 15.55705                                                 fv = 10.19698                                                 fv = 8.753399                                                 fv = 6.386378                                                 fv = 6.386378                                                 fv = 9.097767                                                 fv = 1.340781e+154                                                 fv = 11.76265                                                 fv = 6.999611                                                 fv = 6.999611                                                 fv = 1.340781e+154                                                 fv = 8.175627                                                 fv = 4231840                                                 fv = 1.340781e+154                                                 fv = 82.0237                                                 fv = 1.340781e+154                                                 fv = 13.25153                                                 fv = 8.770495                                                 fv = 10.59034                                                 fv = 9.773011                                                 fv = 7.586966                                                 fv = 1.340781e+154                                                 fv = 25.87042                                                 fv = 8.371015                                                 fv = 6.694731                                                 fv = 93.19903                                                 fv = 1.340781e+154                                                 fv = 6.386378                                                 fv = 23.62661                                                 fv = 6.386378                                                 fv = 6.386378                                                 fv = 6.386378                                                 fv = 6.386378                                                 fv = 1.340781e+154                                                 fv = 6.915562                                                 fv = 11.65211                                                 fv = 1.340781e+154                                                 fv = 13.16691                                                 fv = 9.324071                                                 fv = 6.386378                                                 fv = 6.386378                                                 fv = 1.340781e+154                                                 fv = 6.386378                                                 fv = 34.81202                                                 fv = 8.778489 run # 4: f=6.386378356
#> 
#> bb eval : 633
#> best    : 6.386378356
#> worst   : 1.340780793e+154
#> solution: x = ( 1 0 0 0 0 0 2 0 1 5 1 1 1 1 1 4 1 1 1 1 ) f(x) = 6.386378356
#> 
#>                                                 fv = 6.386378                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (3): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 21.39419